Amazonka.MediaPackageVOD.Types

Description

Service Configuration

defaultService :: Service

API version 2018-11-07 of the Amazon Elemental MediaPackage VOD SDK configuration.

Errors

_UnprocessableEntityException :: AsError a => Getting (First ServiceError) a ServiceError

The parameters sent in the request are not valid.

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

The client is not authorized to access the requested resource.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource does not exist.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

The client has exceeded their resource or throttling limits.

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error occurred.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error occurred.

AdMarkers

newtype AdMarkers

Constructors

AdMarkers'

Fields

fromAdMarkers :: Text

Bundled Patterns

pattern AdMarkers_NONE :: AdMarkers

pattern AdMarkers_PASSTHROUGH :: AdMarkers

pattern AdMarkers_SCTE35_ENHANCED :: AdMarkers

EncryptionMethod

newtype EncryptionMethod

Constructors

EncryptionMethod'

Fields

fromEncryptionMethod :: Text

Bundled Patterns

pattern EncryptionMethod_AES_128 :: EncryptionMethod

pattern EncryptionMethod_SAMPLE_AES :: EncryptionMethod

ManifestLayout

newtype ManifestLayout

Constructors

ManifestLayout'

Fields

fromManifestLayout :: Text

Bundled Patterns

pattern ManifestLayout_COMPACT :: ManifestLayout

pattern ManifestLayout_FULL :: ManifestLayout

PeriodTriggersElement

newtype PeriodTriggersElement

Constructors

PeriodTriggersElement'

Fields

fromPeriodTriggersElement :: Text

Bundled Patterns

pattern PeriodTriggersElement_ADS :: PeriodTriggersElement

Profile

newtype Profile

Constructors

Profile'

Fields

fromProfile :: Text

Bundled Patterns

pattern Profile_HBBTV_1_5 :: Profile

pattern Profile_NONE :: Profile

SegmentTemplateFormat

newtype SegmentTemplateFormat

Constructors

SegmentTemplateFormat'

Fields

fromSegmentTemplateFormat :: Text

Bundled Patterns

pattern SegmentTemplateFormat_NUMBER_WITH_DURATION :: SegmentTemplateFormat

pattern SegmentTemplateFormat_NUMBER_WITH_TIMELINE :: SegmentTemplateFormat

pattern SegmentTemplateFormat_TIME_WITH_TIMELINE :: SegmentTemplateFormat

StreamOrder

newtype StreamOrder

Constructors

StreamOrder'

Fields

fromStreamOrder :: Text

Bundled Patterns

pattern StreamOrder_ORIGINAL :: StreamOrder

pattern StreamOrder_VIDEO_BITRATE_ASCENDING :: StreamOrder

pattern StreamOrder_VIDEO_BITRATE_DESCENDING :: StreamOrder

AssetShallow

data AssetShallow

A MediaPackage VOD Asset resource.

See: newAssetShallow smart constructor.

Constructors

AssetShallow'

Fields

resourceId :: Maybe Text

The resource ID to include in SPEKE key requests.

arn :: Maybe Text

The ARN of the Asset.

createdAt :: Maybe Text

The time the Asset was initially submitted for Ingest.

packagingGroupId :: Maybe Text

The ID of the PackagingGroup for the Asset.

sourceArn :: Maybe Text

ARN of the source object in S3.

sourceRoleArn :: Maybe Text

The IAM role ARN used to access the source S3 bucket.

id :: Maybe Text

The unique identifier for the Asset.

tags :: Maybe (HashMap Text Text)

newAssetShallow :: AssetShallow

Create a value of AssetShallow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:AssetShallow', assetShallow_resourceId - The resource ID to include in SPEKE key requests.

$sel:arn:AssetShallow', assetShallow_arn - The ARN of the Asset.

$sel:createdAt:AssetShallow', assetShallow_createdAt - The time the Asset was initially submitted for Ingest.

$sel:packagingGroupId:AssetShallow', assetShallow_packagingGroupId - The ID of the PackagingGroup for the Asset.

$sel:sourceArn:AssetShallow', assetShallow_sourceArn - ARN of the source object in S3.

$sel:sourceRoleArn:AssetShallow', assetShallow_sourceRoleArn - The IAM role ARN used to access the source S3 bucket.

$sel:id:AssetShallow', assetShallow_id - The unique identifier for the Asset.

$sel:tags:AssetShallow', assetShallow_tags - Undocumented member.

assetShallow_resourceId :: Lens' AssetShallow (Maybe Text)

The resource ID to include in SPEKE key requests.

assetShallow_arn :: Lens' AssetShallow (Maybe Text)

The ARN of the Asset.

assetShallow_createdAt :: Lens' AssetShallow (Maybe Text)

The time the Asset was initially submitted for Ingest.

assetShallow_packagingGroupId :: Lens' AssetShallow (Maybe Text)

The ID of the PackagingGroup for the Asset.

assetShallow_sourceArn :: Lens' AssetShallow (Maybe Text)

ARN of the source object in S3.

assetShallow_sourceRoleArn :: Lens' AssetShallow (Maybe Text)

The IAM role ARN used to access the source S3 bucket.

assetShallow_id :: Lens' AssetShallow (Maybe Text)

The unique identifier for the Asset.

assetShallow_tags :: Lens' AssetShallow (Maybe (HashMap Text Text))

Undocumented member.

Authorization

data Authorization

CDN Authorization credentials

See: newAuthorization smart constructor.

Constructors

Authorization'

Fields

secretsRoleArn :: Text

The Amazon Resource Name (ARN) for the IAM role that allows MediaPackage to communicate with AWS Secrets Manager.

cdnIdentifierSecret :: Text

The Amazon Resource Name (ARN) for the secret in AWS Secrets Manager that is used for CDN authorization.

newAuthorization

Arguments

:: Text
$sel:secretsRoleArn:Authorization'

-> Text
$sel:cdnIdentifierSecret:Authorization'

-> Authorization

Create a value of Authorization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secretsRoleArn:Authorization', authorization_secretsRoleArn - The Amazon Resource Name (ARN) for the IAM role that allows MediaPackage to communicate with AWS Secrets Manager.

$sel:cdnIdentifierSecret:Authorization', authorization_cdnIdentifierSecret - The Amazon Resource Name (ARN) for the secret in AWS Secrets Manager that is used for CDN authorization.

authorization_secretsRoleArn :: Lens' Authorization Text

The Amazon Resource Name (ARN) for the IAM role that allows MediaPackage to communicate with AWS Secrets Manager.

authorization_cdnIdentifierSecret :: Lens' Authorization Text

The Amazon Resource Name (ARN) for the secret in AWS Secrets Manager that is used for CDN authorization.

CmafEncryption

data CmafEncryption

A CMAF encryption configuration.

See: newCmafEncryption smart constructor.

Constructors

CmafEncryption'

Fields

constantInitializationVector :: Maybe Text

An optional 128-bit, 16-byte hex value represented by a 32-character string, used in conjunction with the key for encrypting blocks. If you don't specify a value, then MediaPackage creates the constant initialization vector (IV).

spekeKeyProvider :: SpekeKeyProvider

newCmafEncryption

Arguments

:: SpekeKeyProvider
$sel:spekeKeyProvider:CmafEncryption'

-> CmafEncryption

Create a value of CmafEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:constantInitializationVector:CmafEncryption', cmafEncryption_constantInitializationVector - An optional 128-bit, 16-byte hex value represented by a 32-character string, used in conjunction with the key for encrypting blocks. If you don't specify a value, then MediaPackage creates the constant initialization vector (IV).

$sel:spekeKeyProvider:CmafEncryption', cmafEncryption_spekeKeyProvider - Undocumented member.

cmafEncryption_constantInitializationVector :: Lens' CmafEncryption (Maybe Text)

An optional 128-bit, 16-byte hex value represented by a 32-character string, used in conjunction with the key for encrypting blocks. If you don't specify a value, then MediaPackage creates the constant initialization vector (IV).

cmafEncryption_spekeKeyProvider :: Lens' CmafEncryption SpekeKeyProvider

Undocumented member.

CmafPackage

data CmafPackage

A CMAF packaging configuration.

See: newCmafPackage smart constructor.

Constructors

CmafPackage'

Fields

includeEncoderConfigurationInSegments :: Maybe Bool

When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

segmentDurationSeconds :: Maybe Int

Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

encryption :: Maybe CmafEncryption

hlsManifests :: [HlsManifest]

A list of HLS manifest configurations.

newCmafPackage :: CmafPackage

Create a value of CmafPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeEncoderConfigurationInSegments:CmafPackage', cmafPackage_includeEncoderConfigurationInSegments - When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

$sel:segmentDurationSeconds:CmafPackage', cmafPackage_segmentDurationSeconds - Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

$sel:encryption:CmafPackage', cmafPackage_encryption - Undocumented member.

$sel:hlsManifests:CmafPackage', cmafPackage_hlsManifests - A list of HLS manifest configurations.

cmafPackage_includeEncoderConfigurationInSegments :: Lens' CmafPackage (Maybe Bool)

When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

cmafPackage_segmentDurationSeconds :: Lens' CmafPackage (Maybe Int)

Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

cmafPackage_encryption :: Lens' CmafPackage (Maybe CmafEncryption)

Undocumented member.

cmafPackage_hlsManifests :: Lens' CmafPackage [HlsManifest]

A list of HLS manifest configurations.

DashEncryption

data DashEncryption

A Dynamic Adaptive Streaming over HTTP (DASH) encryption configuration.

See: newDashEncryption smart constructor.

Constructors

DashEncryption'

Fields

spekeKeyProvider :: SpekeKeyProvider

newDashEncryption

Arguments

:: SpekeKeyProvider
$sel:spekeKeyProvider:DashEncryption'

-> DashEncryption

Create a value of DashEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:spekeKeyProvider:DashEncryption', dashEncryption_spekeKeyProvider - Undocumented member.

dashEncryption_spekeKeyProvider :: Lens' DashEncryption SpekeKeyProvider

Undocumented member.

DashManifest

data DashManifest

A DASH manifest configuration.

See: newDashManifest smart constructor.

Constructors

DashManifest'

Fields

minBufferTimeSeconds :: Maybe Int

Minimum duration (in seconds) that a player will buffer media before starting the presentation.

manifestName :: Maybe Text

An optional string to include in the name of the manifest.

profile :: Maybe Profile

The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to "HBBTV_1_5", HbbTV 1.5 compliant output is enabled.

streamSelection :: Maybe StreamSelection

manifestLayout :: Maybe ManifestLayout

Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.

newDashManifest :: DashManifest

Create a value of DashManifest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minBufferTimeSeconds:DashManifest', dashManifest_minBufferTimeSeconds - Minimum duration (in seconds) that a player will buffer media before starting the presentation.

$sel:manifestName:DashManifest', dashManifest_manifestName - An optional string to include in the name of the manifest.

$sel:profile:DashManifest', dashManifest_profile - The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to "HBBTV_1_5", HbbTV 1.5 compliant output is enabled.

$sel:streamSelection:DashManifest', dashManifest_streamSelection - Undocumented member.

$sel:manifestLayout:DashManifest', dashManifest_manifestLayout - Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.

dashManifest_minBufferTimeSeconds :: Lens' DashManifest (Maybe Int)

Minimum duration (in seconds) that a player will buffer media before starting the presentation.

dashManifest_manifestName :: Lens' DashManifest (Maybe Text)

An optional string to include in the name of the manifest.

dashManifest_profile :: Lens' DashManifest (Maybe Profile)

The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to "HBBTV_1_5", HbbTV 1.5 compliant output is enabled.

dashManifest_streamSelection :: Lens' DashManifest (Maybe StreamSelection)

Undocumented member.

dashManifest_manifestLayout :: Lens' DashManifest (Maybe ManifestLayout)

Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.

DashPackage

data DashPackage

A Dynamic Adaptive Streaming over HTTP (DASH) packaging configuration.

See: newDashPackage smart constructor.

Constructors

DashPackage'

Fields

includeEncoderConfigurationInSegments :: Maybe Bool

When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

segmentTemplateFormat :: Maybe SegmentTemplateFormat

Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

segmentDurationSeconds :: Maybe Int

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

encryption :: Maybe DashEncryption

periodTriggers :: Maybe [PeriodTriggersElement]

A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad markers.

dashManifests :: [DashManifest]

A list of DASH manifest configurations.

newDashPackage :: DashPackage

Create a value of DashPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeEncoderConfigurationInSegments:DashPackage', dashPackage_includeEncoderConfigurationInSegments - When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

$sel:segmentTemplateFormat:DashPackage', dashPackage_segmentTemplateFormat - Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

$sel:segmentDurationSeconds:DashPackage', dashPackage_segmentDurationSeconds - Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

$sel:encryption:DashPackage', dashPackage_encryption - Undocumented member.

$sel:periodTriggers:DashPackage', dashPackage_periodTriggers - A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad markers.

$sel:dashManifests:DashPackage', dashPackage_dashManifests - A list of DASH manifest configurations.

dashPackage_includeEncoderConfigurationInSegments :: Lens' DashPackage (Maybe Bool)

When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

dashPackage_segmentTemplateFormat :: Lens' DashPackage (Maybe SegmentTemplateFormat)

Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

dashPackage_segmentDurationSeconds :: Lens' DashPackage (Maybe Int)

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

dashPackage_encryption :: Lens' DashPackage (Maybe DashEncryption)

Undocumented member.

dashPackage_periodTriggers :: Lens' DashPackage (Maybe [PeriodTriggersElement])

A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad markers.

dashPackage_dashManifests :: Lens' DashPackage [DashManifest]

A list of DASH manifest configurations.

EgressAccessLogs

data EgressAccessLogs

Configure egress access logging.

See: newEgressAccessLogs smart constructor.

Constructors

EgressAccessLogs'

Fields

logGroupName :: Maybe Text

Customize the log group name.

newEgressAccessLogs :: EgressAccessLogs

Create a value of EgressAccessLogs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logGroupName:EgressAccessLogs', egressAccessLogs_logGroupName - Customize the log group name.

egressAccessLogs_logGroupName :: Lens' EgressAccessLogs (Maybe Text)

Customize the log group name.

EgressEndpoint

data EgressEndpoint

The endpoint URL used to access an Asset using one PackagingConfiguration.

See: newEgressEndpoint smart constructor.

Constructors

EgressEndpoint'

Fields

status :: Maybe Text

The current processing status of the asset used for the packaging configuration. The status can be either QUEUED, PROCESSING, PLAYABLE, or FAILED. Status information won't be available for most assets ingested before 2021-09-30.

url :: Maybe Text

The URL of the parent manifest for the repackaged Asset.

packagingConfigurationId :: Maybe Text

The ID of the PackagingConfiguration being applied to the Asset.

newEgressEndpoint :: EgressEndpoint

Create a value of EgressEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:EgressEndpoint', egressEndpoint_status - The current processing status of the asset used for the packaging configuration. The status can be either QUEUED, PROCESSING, PLAYABLE, or FAILED. Status information won't be available for most assets ingested before 2021-09-30.

$sel:url:EgressEndpoint', egressEndpoint_url - The URL of the parent manifest for the repackaged Asset.

$sel:packagingConfigurationId:EgressEndpoint', egressEndpoint_packagingConfigurationId - The ID of the PackagingConfiguration being applied to the Asset.

egressEndpoint_status :: Lens' EgressEndpoint (Maybe Text)

The current processing status of the asset used for the packaging configuration. The status can be either QUEUED, PROCESSING, PLAYABLE, or FAILED. Status information won't be available for most assets ingested before 2021-09-30.

egressEndpoint_url :: Lens' EgressEndpoint (Maybe Text)

The URL of the parent manifest for the repackaged Asset.

egressEndpoint_packagingConfigurationId :: Lens' EgressEndpoint (Maybe Text)

The ID of the PackagingConfiguration being applied to the Asset.

HlsEncryption

data HlsEncryption

An HTTP Live Streaming (HLS) encryption configuration.

See: newHlsEncryption smart constructor.

Constructors

HlsEncryption'

Fields

encryptionMethod :: Maybe EncryptionMethod

The encryption method to use.

constantInitializationVector :: Maybe Text

A constant initialization vector for encryption (optional). When not specified the initialization vector will be periodically rotated.

spekeKeyProvider :: SpekeKeyProvider

newHlsEncryption

Arguments

:: SpekeKeyProvider
$sel:spekeKeyProvider:HlsEncryption'

-> HlsEncryption

Create a value of HlsEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionMethod:HlsEncryption', hlsEncryption_encryptionMethod - The encryption method to use.

$sel:constantInitializationVector:HlsEncryption', hlsEncryption_constantInitializationVector - A constant initialization vector for encryption (optional). When not specified the initialization vector will be periodically rotated.

$sel:spekeKeyProvider:HlsEncryption', hlsEncryption_spekeKeyProvider - Undocumented member.

hlsEncryption_encryptionMethod :: Lens' HlsEncryption (Maybe EncryptionMethod)

The encryption method to use.

hlsEncryption_constantInitializationVector :: Lens' HlsEncryption (Maybe Text)

A constant initialization vector for encryption (optional). When not specified the initialization vector will be periodically rotated.

hlsEncryption_spekeKeyProvider :: Lens' HlsEncryption SpekeKeyProvider

Undocumented member.

HlsManifest

data HlsManifest

An HTTP Live Streaming (HLS) manifest configuration.

See: newHlsManifest smart constructor.

Constructors

HlsManifest'

Fields

manifestName :: Maybe Text

An optional string to include in the name of the manifest.

programDateTimeIntervalSeconds :: Maybe Int

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

streamSelection :: Maybe StreamSelection

adMarkers :: Maybe AdMarkers

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source.

includeIframeOnlyStream :: Maybe Bool

When enabled, an I-Frame only stream will be included in the output.

repeatExtXKey :: Maybe Bool

When enabled, the EXT-X-KEY tag will be repeated in output manifests.

newHlsManifest :: HlsManifest

Create a value of HlsManifest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manifestName:HlsManifest', hlsManifest_manifestName - An optional string to include in the name of the manifest.

$sel:programDateTimeIntervalSeconds:HlsManifest', hlsManifest_programDateTimeIntervalSeconds - The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

$sel:streamSelection:HlsManifest', hlsManifest_streamSelection - Undocumented member.

$sel:adMarkers:HlsManifest', hlsManifest_adMarkers - This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source.

$sel:includeIframeOnlyStream:HlsManifest', hlsManifest_includeIframeOnlyStream - When enabled, an I-Frame only stream will be included in the output.

$sel:repeatExtXKey:HlsManifest', hlsManifest_repeatExtXKey - When enabled, the EXT-X-KEY tag will be repeated in output manifests.

hlsManifest_manifestName :: Lens' HlsManifest (Maybe Text)

An optional string to include in the name of the manifest.

hlsManifest_programDateTimeIntervalSeconds :: Lens' HlsManifest (Maybe Int)

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

hlsManifest_streamSelection :: Lens' HlsManifest (Maybe StreamSelection)

Undocumented member.

hlsManifest_adMarkers :: Lens' HlsManifest (Maybe AdMarkers)

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source.

hlsManifest_includeIframeOnlyStream :: Lens' HlsManifest (Maybe Bool)

When enabled, an I-Frame only stream will be included in the output.

hlsManifest_repeatExtXKey :: Lens' HlsManifest (Maybe Bool)

When enabled, the EXT-X-KEY tag will be repeated in output manifests.

HlsPackage

data HlsPackage

An HTTP Live Streaming (HLS) packaging configuration.

See: newHlsPackage smart constructor.

Constructors

HlsPackage'

Fields

useAudioRenditionGroup :: Maybe Bool

When enabled, audio streams will be placed in rendition groups in the output.

includeDvbSubtitles :: Maybe Bool

When enabled, MediaPackage passes through digital video broadcasting (DVB) subtitles into the output.

segmentDurationSeconds :: Maybe Int

Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

encryption :: Maybe HlsEncryption

hlsManifests :: [HlsManifest]

A list of HLS manifest configurations.

newHlsPackage :: HlsPackage

Create a value of HlsPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:useAudioRenditionGroup:HlsPackage', hlsPackage_useAudioRenditionGroup - When enabled, audio streams will be placed in rendition groups in the output.

$sel:includeDvbSubtitles:HlsPackage', hlsPackage_includeDvbSubtitles - When enabled, MediaPackage passes through digital video broadcasting (DVB) subtitles into the output.

$sel:segmentDurationSeconds:HlsPackage', hlsPackage_segmentDurationSeconds - Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

$sel:encryption:HlsPackage', hlsPackage_encryption - Undocumented member.

$sel:hlsManifests:HlsPackage', hlsPackage_hlsManifests - A list of HLS manifest configurations.

hlsPackage_useAudioRenditionGroup :: Lens' HlsPackage (Maybe Bool)

When enabled, audio streams will be placed in rendition groups in the output.

hlsPackage_includeDvbSubtitles :: Lens' HlsPackage (Maybe Bool)

When enabled, MediaPackage passes through digital video broadcasting (DVB) subtitles into the output.

hlsPackage_segmentDurationSeconds :: Lens' HlsPackage (Maybe Int)

Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

hlsPackage_encryption :: Lens' HlsPackage (Maybe HlsEncryption)

Undocumented member.

hlsPackage_hlsManifests :: Lens' HlsPackage [HlsManifest]

A list of HLS manifest configurations.

MssEncryption

data MssEncryption

A Microsoft Smooth Streaming (MSS) encryption configuration.

See: newMssEncryption smart constructor.

Constructors

MssEncryption'

Fields

spekeKeyProvider :: SpekeKeyProvider

newMssEncryption

Arguments

:: SpekeKeyProvider
$sel:spekeKeyProvider:MssEncryption'

-> MssEncryption

Create a value of MssEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:spekeKeyProvider:MssEncryption', mssEncryption_spekeKeyProvider - Undocumented member.

mssEncryption_spekeKeyProvider :: Lens' MssEncryption SpekeKeyProvider

Undocumented member.

MssManifest

data MssManifest

A Microsoft Smooth Streaming (MSS) manifest configuration.

See: newMssManifest smart constructor.

Constructors

MssManifest'

Fields

manifestName :: Maybe Text

An optional string to include in the name of the manifest.

streamSelection :: Maybe StreamSelection

newMssManifest :: MssManifest

Create a value of MssManifest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manifestName:MssManifest', mssManifest_manifestName - An optional string to include in the name of the manifest.

$sel:streamSelection:MssManifest', mssManifest_streamSelection - Undocumented member.

mssManifest_manifestName :: Lens' MssManifest (Maybe Text)

An optional string to include in the name of the manifest.

mssManifest_streamSelection :: Lens' MssManifest (Maybe StreamSelection)

Undocumented member.

MssPackage

data MssPackage

A Microsoft Smooth Streaming (MSS) PackagingConfiguration.

See: newMssPackage smart constructor.

Constructors

MssPackage'

Fields

segmentDurationSeconds :: Maybe Int

The duration (in seconds) of each segment.

encryption :: Maybe MssEncryption

mssManifests :: [MssManifest]

A list of MSS manifest configurations.

newMssPackage :: MssPackage

Create a value of MssPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:segmentDurationSeconds:MssPackage', mssPackage_segmentDurationSeconds - The duration (in seconds) of each segment.

$sel:encryption:MssPackage', mssPackage_encryption - Undocumented member.

$sel:mssManifests:MssPackage', mssPackage_mssManifests - A list of MSS manifest configurations.

mssPackage_segmentDurationSeconds :: Lens' MssPackage (Maybe Int)

The duration (in seconds) of each segment.

mssPackage_encryption :: Lens' MssPackage (Maybe MssEncryption)

Undocumented member.

mssPackage_mssManifests :: Lens' MssPackage [MssManifest]

A list of MSS manifest configurations.

PackagingConfiguration

data PackagingConfiguration

A MediaPackage VOD PackagingConfiguration resource.

See: newPackagingConfiguration smart constructor.

Constructors

PackagingConfiguration'

Fields

hlsPackage :: Maybe HlsPackage

arn :: Maybe Text

The ARN of the PackagingConfiguration.

packagingGroupId :: Maybe Text

The ID of a PackagingGroup.

dashPackage :: Maybe DashPackage

mssPackage :: Maybe MssPackage

id :: Maybe Text

The ID of the PackagingConfiguration.

cmafPackage :: Maybe CmafPackage

tags :: Maybe (HashMap Text Text)

newPackagingConfiguration :: PackagingConfiguration

Create a value of PackagingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hlsPackage:PackagingConfiguration', packagingConfiguration_hlsPackage - Undocumented member.

$sel:arn:PackagingConfiguration', packagingConfiguration_arn - The ARN of the PackagingConfiguration.

$sel:packagingGroupId:PackagingConfiguration', packagingConfiguration_packagingGroupId - The ID of a PackagingGroup.

$sel:dashPackage:PackagingConfiguration', packagingConfiguration_dashPackage - Undocumented member.

$sel:mssPackage:PackagingConfiguration', packagingConfiguration_mssPackage - Undocumented member.

$sel:id:PackagingConfiguration', packagingConfiguration_id - The ID of the PackagingConfiguration.

$sel:cmafPackage:PackagingConfiguration', packagingConfiguration_cmafPackage - Undocumented member.

$sel:tags:PackagingConfiguration', packagingConfiguration_tags - Undocumented member.

packagingConfiguration_hlsPackage :: Lens' PackagingConfiguration (Maybe HlsPackage)

Undocumented member.

packagingConfiguration_arn :: Lens' PackagingConfiguration (Maybe Text)

The ARN of the PackagingConfiguration.

packagingConfiguration_packagingGroupId :: Lens' PackagingConfiguration (Maybe Text)

The ID of a PackagingGroup.

packagingConfiguration_dashPackage :: Lens' PackagingConfiguration (Maybe DashPackage)

Undocumented member.

packagingConfiguration_mssPackage :: Lens' PackagingConfiguration (Maybe MssPackage)

Undocumented member.

packagingConfiguration_id :: Lens' PackagingConfiguration (Maybe Text)

The ID of the PackagingConfiguration.

packagingConfiguration_cmafPackage :: Lens' PackagingConfiguration (Maybe CmafPackage)

Undocumented member.

packagingConfiguration_tags :: Lens' PackagingConfiguration (Maybe (HashMap Text Text))

Undocumented member.

PackagingGroup

data PackagingGroup

A MediaPackage VOD PackagingGroup resource.

See: newPackagingGroup smart constructor.

Constructors

PackagingGroup'

Fields

arn :: Maybe Text

The ARN of the PackagingGroup.

authorization :: Maybe Authorization

domainName :: Maybe Text

The fully qualified domain name for Assets in the PackagingGroup.

id :: Maybe Text

The ID of the PackagingGroup.

egressAccessLogs :: Maybe EgressAccessLogs

tags :: Maybe (HashMap Text Text)

newPackagingGroup :: PackagingGroup

Create a value of PackagingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:PackagingGroup', packagingGroup_arn - The ARN of the PackagingGroup.

$sel:authorization:PackagingGroup', packagingGroup_authorization - Undocumented member.

$sel:domainName:PackagingGroup', packagingGroup_domainName - The fully qualified domain name for Assets in the PackagingGroup.

$sel:id:PackagingGroup', packagingGroup_id - The ID of the PackagingGroup.

$sel:egressAccessLogs:PackagingGroup', packagingGroup_egressAccessLogs - Undocumented member.

$sel:tags:PackagingGroup', packagingGroup_tags - Undocumented member.

packagingGroup_arn :: Lens' PackagingGroup (Maybe Text)

The ARN of the PackagingGroup.

packagingGroup_authorization :: Lens' PackagingGroup (Maybe Authorization)

Undocumented member.

packagingGroup_domainName :: Lens' PackagingGroup (Maybe Text)

The fully qualified domain name for Assets in the PackagingGroup.

packagingGroup_id :: Lens' PackagingGroup (Maybe Text)

The ID of the PackagingGroup.

packagingGroup_egressAccessLogs :: Lens' PackagingGroup (Maybe EgressAccessLogs)

Undocumented member.

packagingGroup_tags :: Lens' PackagingGroup (Maybe (HashMap Text Text))

Undocumented member.

SpekeKeyProvider

data SpekeKeyProvider

A configuration for accessing an external Secure Packager and Encoder Key Exchange (SPEKE) service that will provide encryption keys.

See: newSpekeKeyProvider smart constructor.

Constructors

SpekeKeyProvider'

Fields

systemIds :: [Text]

The system IDs to include in key requests.

url :: Text

The URL of the external key provider service.

roleArn :: Text

An Amazon Resource Name (ARN) of an IAM role that AWS Elemental MediaPackage will assume when accessing the key provider service.

newSpekeKeyProvider

Arguments

:: Text
$sel:url:SpekeKeyProvider'

-> Text
$sel:roleArn:SpekeKeyProvider'

-> SpekeKeyProvider

Create a value of SpekeKeyProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:systemIds:SpekeKeyProvider', spekeKeyProvider_systemIds - The system IDs to include in key requests.

$sel:url:SpekeKeyProvider', spekeKeyProvider_url - The URL of the external key provider service.

$sel:roleArn:SpekeKeyProvider', spekeKeyProvider_roleArn - An Amazon Resource Name (ARN) of an IAM role that AWS Elemental MediaPackage will assume when accessing the key provider service.

spekeKeyProvider_systemIds :: Lens' SpekeKeyProvider [Text]

The system IDs to include in key requests.

spekeKeyProvider_url :: Lens' SpekeKeyProvider Text

The URL of the external key provider service.

spekeKeyProvider_roleArn :: Lens' SpekeKeyProvider Text

An Amazon Resource Name (ARN) of an IAM role that AWS Elemental MediaPackage will assume when accessing the key provider service.

StreamSelection

data StreamSelection

A StreamSelection configuration.

See: newStreamSelection smart constructor.

Constructors

StreamSelection'

Fields

streamOrder :: Maybe StreamOrder

A directive that determines the order of streams in the output.

minVideoBitsPerSecond :: Maybe Int

The minimum video bitrate (bps) to include in output.

maxVideoBitsPerSecond :: Maybe Int

The maximum video bitrate (bps) to include in output.

newStreamSelection :: StreamSelection

Create a value of StreamSelection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamOrder:StreamSelection', streamSelection_streamOrder - A directive that determines the order of streams in the output.

$sel:minVideoBitsPerSecond:StreamSelection', streamSelection_minVideoBitsPerSecond - The minimum video bitrate (bps) to include in output.

$sel:maxVideoBitsPerSecond:StreamSelection', streamSelection_maxVideoBitsPerSecond - The maximum video bitrate (bps) to include in output.

streamSelection_streamOrder :: Lens' StreamSelection (Maybe StreamOrder)

A directive that determines the order of streams in the output.

streamSelection_minVideoBitsPerSecond :: Lens' StreamSelection (Maybe Int)

The minimum video bitrate (bps) to include in output.

streamSelection_maxVideoBitsPerSecond :: Lens' StreamSelection (Maybe Int)

The maximum video bitrate (bps) to include in output.