Amazonka.MediaPackageVOD.Types.CmafPackage

Description

Documentation

data CmafPackage

A CMAF packaging configuration.

See: newCmafPackage smart constructor.

Constructors

CmafPackage'

Fields

includeEncoderConfigurationInSegments :: Maybe Bool

When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

segmentDurationSeconds :: Maybe Int

Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

encryption :: Maybe CmafEncryption

hlsManifests :: [HlsManifest]

A list of HLS manifest configurations.

newCmafPackage :: CmafPackage

Create a value of CmafPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeEncoderConfigurationInSegments:CmafPackage', cmafPackage_includeEncoderConfigurationInSegments - When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

$sel:segmentDurationSeconds:CmafPackage', cmafPackage_segmentDurationSeconds - Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

$sel:encryption:CmafPackage', cmafPackage_encryption - Undocumented member.

$sel:hlsManifests:CmafPackage', cmafPackage_hlsManifests - A list of HLS manifest configurations.

cmafPackage_includeEncoderConfigurationInSegments :: Lens' CmafPackage (Maybe Bool)

When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

cmafPackage_segmentDurationSeconds :: Lens' CmafPackage (Maybe Int)

Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

cmafPackage_encryption :: Lens' CmafPackage (Maybe CmafEncryption)

Undocumented member.

cmafPackage_hlsManifests :: Lens' CmafPackage [HlsManifest]

A list of HLS manifest configurations.