Amazonka.MediaPackageVOD.Types.HlsManifest

Description

Documentation

data HlsManifest

An HTTP Live Streaming (HLS) manifest configuration.

See: newHlsManifest smart constructor.

Constructors

HlsManifest'

Fields

manifestName :: Maybe Text

An optional string to include in the name of the manifest.

programDateTimeIntervalSeconds :: Maybe Int

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

streamSelection :: Maybe StreamSelection

adMarkers :: Maybe AdMarkers

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source.

includeIframeOnlyStream :: Maybe Bool

When enabled, an I-Frame only stream will be included in the output.

repeatExtXKey :: Maybe Bool

When enabled, the EXT-X-KEY tag will be repeated in output manifests.

newHlsManifest :: HlsManifest

Create a value of HlsManifest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manifestName:HlsManifest', hlsManifest_manifestName - An optional string to include in the name of the manifest.

$sel:programDateTimeIntervalSeconds:HlsManifest', hlsManifest_programDateTimeIntervalSeconds - The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

$sel:streamSelection:HlsManifest', hlsManifest_streamSelection - Undocumented member.

$sel:adMarkers:HlsManifest', hlsManifest_adMarkers - This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source.

$sel:includeIframeOnlyStream:HlsManifest', hlsManifest_includeIframeOnlyStream - When enabled, an I-Frame only stream will be included in the output.

$sel:repeatExtXKey:HlsManifest', hlsManifest_repeatExtXKey - When enabled, the EXT-X-KEY tag will be repeated in output manifests.

hlsManifest_manifestName :: Lens' HlsManifest (Maybe Text)

An optional string to include in the name of the manifest.

hlsManifest_programDateTimeIntervalSeconds :: Lens' HlsManifest (Maybe Int)

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

hlsManifest_streamSelection :: Lens' HlsManifest (Maybe StreamSelection)

Undocumented member.

hlsManifest_adMarkers :: Lens' HlsManifest (Maybe AdMarkers)

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source.

hlsManifest_includeIframeOnlyStream :: Lens' HlsManifest (Maybe Bool)

When enabled, an I-Frame only stream will be included in the output.

hlsManifest_repeatExtXKey :: Lens' HlsManifest (Maybe Bool)

When enabled, the EXT-X-KEY tag will be repeated in output manifests.