Amazonka.MediaStore.DescribeContainer

Description

Retrieves the properties of the requested container. This request is commonly used to retrieve the endpoint of a container. An endpoint is a value assigned by the service when a new container is created. A container's endpoint does not change after it has been assigned. The DescribeContainer request returns a single Container object based on ContainerName. To return all Container objects that are associated with a specified AWS account, use ListContainers.

Creating a Request

data DescribeContainer

See: newDescribeContainer smart constructor.

Constructors

DescribeContainer'

Fields

containerName :: Maybe Text

The name of the container to query.

newDescribeContainer :: DescribeContainer

Create a value of DescribeContainer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:DescribeContainer', describeContainer_containerName - The name of the container to query.

Request Lenses

describeContainer_containerName :: Lens' DescribeContainer (Maybe Text)

The name of the container to query.

Destructuring the Response

data DescribeContainerResponse

See: newDescribeContainerResponse smart constructor.

Constructors

DescribeContainerResponse'

Fields

container :: Maybe Container

The name of the queried container.

httpStatus :: Int

The response's http status code.

newDescribeContainerResponse

Arguments

:: Int
$sel:httpStatus:DescribeContainerResponse'

-> DescribeContainerResponse

Create a value of DescribeContainerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:container:DescribeContainerResponse', describeContainerResponse_container - The name of the queried container.

$sel:httpStatus:DescribeContainerResponse', describeContainerResponse_httpStatus - The response's http status code.

Response Lenses

describeContainerResponse_container :: Lens' DescribeContainerResponse (Maybe Container)

The name of the queried container.

describeContainerResponse_httpStatus :: Lens' DescribeContainerResponse Int

The response's http status code.