Amazonka.MediaStore.GetMetricPolicy

Description

Returns the metric policy for the specified container.

Creating a Request

data GetMetricPolicy

See: newGetMetricPolicy smart constructor.

Constructors

GetMetricPolicy'

Fields

containerName :: Text

The name of the container that is associated with the metric policy.

newGetMetricPolicy

Arguments

:: Text
$sel:containerName:GetMetricPolicy'

-> GetMetricPolicy

Create a value of GetMetricPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:GetMetricPolicy', getMetricPolicy_containerName - The name of the container that is associated with the metric policy.

Request Lenses

getMetricPolicy_containerName :: Lens' GetMetricPolicy Text

The name of the container that is associated with the metric policy.

Destructuring the Response

data GetMetricPolicyResponse

See: newGetMetricPolicyResponse smart constructor.

Constructors

GetMetricPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

metricPolicy :: MetricPolicy

The metric policy that is associated with the specific container.

newGetMetricPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetMetricPolicyResponse'

-> MetricPolicy
$sel:metricPolicy:GetMetricPolicyResponse'

-> GetMetricPolicyResponse

Create a value of GetMetricPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetMetricPolicyResponse', getMetricPolicyResponse_httpStatus - The response's http status code.

$sel:metricPolicy:GetMetricPolicyResponse', getMetricPolicyResponse_metricPolicy - The metric policy that is associated with the specific container.

Response Lenses

getMetricPolicyResponse_httpStatus :: Lens' GetMetricPolicyResponse Int

The response's http status code.

getMetricPolicyResponse_metricPolicy :: Lens' GetMetricPolicyResponse MetricPolicy

The metric policy that is associated with the specific container.