Amazonka.MediaStoreData

Description

Derived from API version 2017-09-01 of the AWS service descriptions, licensed under Apache 2.0.

An AWS Elemental MediaStore asset is an object, similar to an object in the Amazon S3 service. Objects are the fundamental entities that are stored in AWS Elemental MediaStore.

Service Configuration

defaultService :: Service

API version 2017-09-01 of the Amazon Elemental MediaStore Data Plane SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MediaStoreData.

RequestedRangeNotSatisfiableException

_RequestedRangeNotSatisfiableException :: AsError a => Getting (First ServiceError) a ServiceError

The requested content range is not valid.

InternalServerError

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable.

ContainerNotFoundException

_ContainerNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified container was not found for the specified account.

ObjectNotFoundException

_ObjectNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Could not perform an operation on an object that does not exist.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

PutObject

data PutObject

See: newPutObject smart constructor.

Constructors

PutObject' (Maybe StorageClass) (Maybe UploadAvailability) (Maybe Text) (Maybe Text) Text HashedBody

newPutObject

Arguments

:: Text
$sel:path:PutObject'

-> HashedBody
$sel:body:PutObject'

-> PutObject

Create a value of PutObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:storageClass:PutObject', putObject_storageClass - Indicates the storage class of a Put request. Defaults to high-performance temporal storage class, and objects are persisted into durable storage shortly after being received.

$sel:uploadAvailability:PutObject', putObject_uploadAvailability - Indicates the availability of an object while it is still uploading. If the value is set to streaming, the object is available for downloading after some initial buffering but before the object is uploaded completely. If the value is set to standard, the object is available for downloading only when it is uploaded completely. The default value for this header is standard.

To use this header, you must also set the HTTP Transfer-Encoding header to chunked.

$sel:cacheControl:PutObject', putObject_cacheControl - An optional CacheControl header that allows the caller to control the object's cache behavior. Headers can be passed in as specified in the HTTP at https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

Headers with a custom user-defined value are also accepted.

$sel:contentType:PutObject', putObject_contentType - The content type of the object.

$sel:path:PutObject', putObject_path - The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

For example, to upload the file mlaw.avi to the folder path premium\canada in the container movies, enter the path premium/canada/mlaw.avi.

Do not include the container name in this path.

If the path includes any folders that don't exist yet, the service creates them. For example, suppose you have an existing premium/usa subfolder. If you specify premium/canada, the service creates a canada subfolder in the premium folder. You then have two subfolders, usa and canada, in the premium folder.

There is no correlation between the path to the source and the path (folders) in the container in AWS Elemental MediaStore.

For more information about folders and how they exist in a container, see the AWS Elemental MediaStore User Guide.

The file name is the name that is assigned to the file that you upload. The file can have the same name inside and outside of AWS Elemental MediaStore, or it can have the same name. The file name can include or omit an extension.

$sel:body:PutObject', putObject_body - The bytes to be stored.

data PutObjectResponse

See: newPutObjectResponse smart constructor.

Constructors

PutObjectResponse' (Maybe Text) (Maybe StorageClass) (Maybe Text) Int

newPutObjectResponse

Arguments

:: Int
$sel:httpStatus:PutObjectResponse'

-> PutObjectResponse

Create a value of PutObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:PutObjectResponse', putObjectResponse_eTag - Unique identifier of the object in the container.

$sel:storageClass:PutObject', putObjectResponse_storageClass - The storage class where the object was persisted. The class should be “Temporal”.

$sel:contentSHA256:PutObjectResponse', putObjectResponse_contentSHA256 - The SHA256 digest of the object that is persisted.

$sel:httpStatus:PutObjectResponse', putObjectResponse_httpStatus - The response's http status code.

DeleteObject

data DeleteObject

See: newDeleteObject smart constructor.

Constructors

DeleteObject' Text

newDeleteObject

Arguments

:: Text
$sel:path:DeleteObject'

-> DeleteObject

Create a value of DeleteObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:DeleteObject', deleteObject_path - The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

data DeleteObjectResponse

See: newDeleteObjectResponse smart constructor.

Constructors

DeleteObjectResponse' Int

newDeleteObjectResponse

Arguments

:: Int
$sel:httpStatus:DeleteObjectResponse'

-> DeleteObjectResponse

Create a value of DeleteObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteObjectResponse', deleteObjectResponse_httpStatus - The response's http status code.

DescribeObject

data DescribeObject

See: newDescribeObject smart constructor.

Constructors

DescribeObject' Text

newDescribeObject

Arguments

:: Text
$sel:path:DescribeObject'

-> DescribeObject

Create a value of DescribeObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:DescribeObject', describeObject_path - The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

data DescribeObjectResponse

See: newDescribeObjectResponse smart constructor.

Constructors

DescribeObjectResponse' (Maybe Text) (Maybe Natural) (Maybe Text) (Maybe POSIX) (Maybe Text) Int

newDescribeObjectResponse

Arguments

:: Int
$sel:httpStatus:DescribeObjectResponse'

-> DescribeObjectResponse

Create a value of DescribeObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:DescribeObjectResponse', describeObjectResponse_eTag - The ETag that represents a unique instance of the object.

$sel:contentLength:DescribeObjectResponse', describeObjectResponse_contentLength - The length of the object in bytes.

$sel:cacheControl:DescribeObjectResponse', describeObjectResponse_cacheControl - An optional CacheControl header that allows the caller to control the object's cache behavior. Headers can be passed in as specified in the HTTP at https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

Headers with a custom user-defined value are also accepted.

$sel:lastModified:DescribeObjectResponse', describeObjectResponse_lastModified - The date and time that the object was last modified.

$sel:contentType:DescribeObjectResponse', describeObjectResponse_contentType - The content type of the object.

$sel:httpStatus:DescribeObjectResponse', describeObjectResponse_httpStatus - The response's http status code.

GetObject

data GetObject

See: newGetObject smart constructor.

Constructors

GetObject' (Maybe Text) Text

newGetObject

Arguments

:: Text
$sel:path:GetObject'

-> GetObject

Create a value of GetObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:range:GetObject', getObject_range - The range bytes of an object to retrieve. For more information about the Range header, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35. AWS Elemental MediaStore ignores this header for partially uploaded objects that have streaming upload availability.

$sel:path:GetObject', getObject_path - The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

For example, to upload the file mlaw.avi to the folder path premium\canada in the container movies, enter the path premium/canada/mlaw.avi.

Do not include the container name in this path.

If the path includes any folders that don't exist yet, the service creates them. For example, suppose you have an existing premium/usa subfolder. If you specify premium/canada, the service creates a canada subfolder in the premium folder. You then have two subfolders, usa and canada, in the premium folder.

There is no correlation between the path to the source and the path (folders) in the container in AWS Elemental MediaStore.

For more information about folders and how they exist in a container, see the AWS Elemental MediaStore User Guide.

The file name is the name that is assigned to the file that you upload. The file can have the same name inside and outside of AWS Elemental MediaStore, or it can have the same name. The file name can include or omit an extension.

data GetObjectResponse

See: newGetObjectResponse smart constructor.

Constructors

GetObjectResponse' (Maybe Text) (Maybe Natural) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe Text) Int ResponseBody

newGetObjectResponse

Arguments

:: Int
$sel:statusCode:GetObjectResponse'

-> ResponseBody
$sel:body:GetObjectResponse'

-> GetObjectResponse

Create a value of GetObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:GetObjectResponse', getObjectResponse_eTag - The ETag that represents a unique instance of the object.

$sel:contentLength:GetObjectResponse', getObjectResponse_contentLength - The length of the object in bytes.

$sel:cacheControl:GetObjectResponse', getObjectResponse_cacheControl - An optional CacheControl header that allows the caller to control the object's cache behavior. Headers can be passed in as specified in the HTTP spec at https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

Headers with a custom user-defined value are also accepted.

$sel:lastModified:GetObjectResponse', getObjectResponse_lastModified - The date and time that the object was last modified.

$sel:contentRange:GetObjectResponse', getObjectResponse_contentRange - The range of bytes to retrieve.

$sel:contentType:GetObjectResponse', getObjectResponse_contentType - The content type of the object.

$sel:statusCode:GetObjectResponse', getObjectResponse_statusCode - The HTML status code of the request. Status codes ranging from 200 to 299 indicate success. All other status codes indicate the type of error that occurred.

$sel:body:GetObjectResponse', getObjectResponse_body - The bytes of the object.

ListItems (Paginated)

data ListItems

See: newListItems smart constructor.

Constructors

ListItems' (Maybe Text) (Maybe Text) (Maybe Natural)

newListItems :: ListItems

Create a value of ListItems with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:ListItems', listItems_path - The path in the container from which to retrieve items. Format: <folder name>/<folder name>/<file name>

$sel:nextToken:ListItems', listItems_nextToken - The token that identifies which batch of results that you want to see. For example, you submit a ListItems request with MaxResults set at 500. The service returns the first batch of results (up to 500) and a NextToken value. To see the next batch of results, you can submit the ListItems request a second time and specify the NextToken value.

Tokens expire after 15 minutes.

$sel:maxResults:ListItems', listItems_maxResults - The maximum number of results to return per API request. For example, you submit a ListItems request with MaxResults set at 500. Although 2,000 items match your request, the service returns no more than the first 500 items. (The service also returns a NextToken value that you can use to fetch the next batch of results.) The service might return fewer results than the MaxResults value.

If MaxResults is not included in the request, the service defaults to pagination with a maximum of 1,000 results per page.

data ListItemsResponse

See: newListItemsResponse smart constructor.

Constructors

ListItemsResponse' (Maybe [Item]) (Maybe Text) Int

newListItemsResponse

Arguments

:: Int
$sel:httpStatus:ListItemsResponse'

-> ListItemsResponse

Create a value of ListItemsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:ListItemsResponse', listItemsResponse_items - The metadata entries for the folders and objects at the requested path.

$sel:nextToken:ListItems', listItemsResponse_nextToken - The token that can be used in a request to view the next set of results. For example, you submit a ListItems request that matches 2,000 items with MaxResults set at 500. The service returns the first batch of results (up to 500) and a NextToken value that can be used to fetch the next batch of results.

$sel:httpStatus:ListItemsResponse', listItemsResponse_httpStatus - The response's http status code.

Types

ItemType

newtype ItemType

Constructors

ItemType'

Fields

fromItemType :: Text

Bundled Patterns

pattern ItemType_FOLDER :: ItemType

pattern ItemType_OBJECT :: ItemType

StorageClass

newtype StorageClass

Constructors

StorageClass'

Fields

fromStorageClass :: Text

Bundled Patterns

pattern StorageClass_TEMPORAL :: StorageClass

UploadAvailability

newtype UploadAvailability

Constructors

UploadAvailability'

Fields

fromUploadAvailability :: Text

Bundled Patterns

pattern UploadAvailability_STANDARD :: UploadAvailability

pattern UploadAvailability_STREAMING :: UploadAvailability

Item

data Item

A metadata entry for a folder or object.

See: newItem smart constructor.

Constructors

Item' (Maybe Text) (Maybe Natural) (Maybe Text) (Maybe ItemType) (Maybe POSIX) (Maybe Text)

newItem :: Item

Create a value of Item with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:Item', item_eTag - The ETag that represents a unique instance of the item.

$sel:contentLength:Item', item_contentLength - The length of the item in bytes.

$sel:name:Item', item_name - The name of the item.

$sel:type':Item', item_type - The item type (folder or object).

$sel:lastModified:Item', item_lastModified - The date and time that the item was last modified.

$sel:contentType:Item', item_contentType - The content type of the item.