Amazonka.MediaTailor.CreateChannel

Description

Creates a channel.

Creating a Request

data CreateChannel

See: newCreateChannel smart constructor.

Constructors

CreateChannel'

Fields

fillerSlate :: Maybe SlateSource

The slate used to fill gaps between programs in the schedule. You must configure filler slate if your channel uses a LINEAR PlaybackMode.

tags :: Maybe (HashMap Text Text)

The tags to assign to the channel.

channelName :: Text

The identifier for the channel you are working on.

outputs :: [RequestOutputItem]

The channel's output properties.

playbackMode :: PlaybackMode

The type of playback mode to use for this channel.

LINEAR - The programs in the schedule play once back-to-back in the schedule.

LOOP - The programs in the schedule play back-to-back in an endless loop. When the last program in the schedule stops playing, playback loops back to the first program in the schedule.

newCreateChannel

Arguments

:: Text
$sel:channelName:CreateChannel'

-> PlaybackMode
$sel:playbackMode:CreateChannel'

-> CreateChannel

Create a value of CreateChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fillerSlate:CreateChannel', createChannel_fillerSlate - The slate used to fill gaps between programs in the schedule. You must configure filler slate if your channel uses a LINEAR PlaybackMode.

$sel:tags:CreateChannel', createChannel_tags - The tags to assign to the channel.

$sel:channelName:CreateChannel', createChannel_channelName - The identifier for the channel you are working on.

$sel:outputs:CreateChannel', createChannel_outputs - The channel's output properties.

$sel:playbackMode:CreateChannel', createChannel_playbackMode - The type of playback mode to use for this channel.

LINEAR - The programs in the schedule play once back-to-back in the schedule.

LOOP - The programs in the schedule play back-to-back in an endless loop. When the last program in the schedule stops playing, playback loops back to the first program in the schedule.

Request Lenses

createChannel_fillerSlate :: Lens' CreateChannel (Maybe SlateSource)

The slate used to fill gaps between programs in the schedule. You must configure filler slate if your channel uses a LINEAR PlaybackMode.

createChannel_tags :: Lens' CreateChannel (Maybe (HashMap Text Text))

The tags to assign to the channel.

createChannel_channelName :: Lens' CreateChannel Text

The identifier for the channel you are working on.

createChannel_outputs :: Lens' CreateChannel [RequestOutputItem]

The channel's output properties.

createChannel_playbackMode :: Lens' CreateChannel PlaybackMode

The type of playback mode to use for this channel.

LINEAR - The programs in the schedule play once back-to-back in the schedule.

LOOP - The programs in the schedule play back-to-back in an endless loop. When the last program in the schedule stops playing, playback loops back to the first program in the schedule.

Destructuring the Response

data CreateChannelResponse

See: newCreateChannelResponse smart constructor.

Constructors

CreateChannelResponse'

Fields

creationTime :: Maybe POSIX

The timestamp of when the channel was created.

arn :: Maybe Text

The ARN of the channel.

lastModifiedTime :: Maybe POSIX

The timestamp of when the channel was last modified.

playbackMode :: Maybe Text

The channel's playback mode.

channelName :: Maybe Text

The name of the channel.

outputs :: Maybe [ResponseOutputItem]

The channel's output properties.

channelState :: Maybe ChannelState

Indicates whether the channel is in a running state or not.

fillerSlate :: Maybe SlateSource

Contains information about the slate used to fill gaps between programs in the schedule.

tags :: Maybe (HashMap Text Text)

The tags assigned to the channel.

httpStatus :: Int

The response's http status code.

newCreateChannelResponse

Arguments

:: Int
$sel:httpStatus:CreateChannelResponse'

-> CreateChannelResponse

Create a value of CreateChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CreateChannelResponse', createChannelResponse_creationTime - The timestamp of when the channel was created.

$sel:arn:CreateChannelResponse', createChannelResponse_arn - The ARN of the channel.

$sel:lastModifiedTime:CreateChannelResponse', createChannelResponse_lastModifiedTime - The timestamp of when the channel was last modified.

$sel:playbackMode:CreateChannel', createChannelResponse_playbackMode - The channel's playback mode.

$sel:channelName:CreateChannel', createChannelResponse_channelName - The name of the channel.

$sel:outputs:CreateChannel', createChannelResponse_outputs - The channel's output properties.

$sel:channelState:CreateChannelResponse', createChannelResponse_channelState - Indicates whether the channel is in a running state or not.

$sel:fillerSlate:CreateChannel', createChannelResponse_fillerSlate - Contains information about the slate used to fill gaps between programs in the schedule.

$sel:tags:CreateChannel', createChannelResponse_tags - The tags assigned to the channel.

$sel:httpStatus:CreateChannelResponse', createChannelResponse_httpStatus - The response's http status code.

Response Lenses

createChannelResponse_creationTime :: Lens' CreateChannelResponse (Maybe UTCTime)

The timestamp of when the channel was created.

createChannelResponse_arn :: Lens' CreateChannelResponse (Maybe Text)

The ARN of the channel.

createChannelResponse_lastModifiedTime :: Lens' CreateChannelResponse (Maybe UTCTime)

The timestamp of when the channel was last modified.

createChannelResponse_playbackMode :: Lens' CreateChannelResponse (Maybe Text)

The channel's playback mode.

createChannelResponse_channelName :: Lens' CreateChannelResponse (Maybe Text)

The name of the channel.

createChannelResponse_outputs :: Lens' CreateChannelResponse (Maybe [ResponseOutputItem])

The channel's output properties.

createChannelResponse_channelState :: Lens' CreateChannelResponse (Maybe ChannelState)

Indicates whether the channel is in a running state or not.

createChannelResponse_fillerSlate :: Lens' CreateChannelResponse (Maybe SlateSource)

Contains information about the slate used to fill gaps between programs in the schedule.

createChannelResponse_tags :: Lens' CreateChannelResponse (Maybe (HashMap Text Text))

The tags assigned to the channel.

createChannelResponse_httpStatus :: Lens' CreateChannelResponse Int

The response's http status code.