Amazonka.MediaTailor.CreatePrefetchSchedule

Description

Creates a new prefetch schedule for the specified playback configuration.

Creating a Request

data CreatePrefetchSchedule

See: newCreatePrefetchSchedule smart constructor.

Constructors

CreatePrefetchSchedule'

Fields

streamId :: Maybe Text

An optional stream identifier that MediaTailor uses to prefetch ads for multiple streams that use the same playback configuration. If StreamId is specified, MediaTailor returns all of the prefetch schedules with an exact match on StreamId. If not specified, MediaTailor returns all of the prefetch schedules for the playback configuration, regardless of StreamId.

name :: Text

The identifier for the playback configuration.

playbackConfigurationName :: Text

The name of the playback configuration.

consumption :: PrefetchConsumption

The configuration settings for MediaTailor's consumption of the prefetched ads from the ad decision server. Each consumption configuration contains an end time and an optional start time that define the consumption window. Prefetch schedules automatically expire no earlier than seven days after the end time.

retrieval :: PrefetchRetrieval

The configuration settings for retrieval of prefetched ads from the ad decision server. Only one set of prefetched ads will be retrieved and subsequently consumed for each ad break.

newCreatePrefetchSchedule

Arguments

:: Text
$sel:name:CreatePrefetchSchedule'

-> Text
$sel:playbackConfigurationName:CreatePrefetchSchedule'

-> PrefetchConsumption
$sel:consumption:CreatePrefetchSchedule'

-> PrefetchRetrieval
$sel:retrieval:CreatePrefetchSchedule'

-> CreatePrefetchSchedule

Create a value of CreatePrefetchSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamId:CreatePrefetchSchedule', createPrefetchSchedule_streamId - An optional stream identifier that MediaTailor uses to prefetch ads for multiple streams that use the same playback configuration. If StreamId is specified, MediaTailor returns all of the prefetch schedules with an exact match on StreamId. If not specified, MediaTailor returns all of the prefetch schedules for the playback configuration, regardless of StreamId.

$sel:name:CreatePrefetchSchedule', createPrefetchSchedule_name - The identifier for the playback configuration.

$sel:playbackConfigurationName:CreatePrefetchSchedule', createPrefetchSchedule_playbackConfigurationName - The name of the playback configuration.

$sel:consumption:CreatePrefetchSchedule', createPrefetchSchedule_consumption - The configuration settings for MediaTailor's consumption of the prefetched ads from the ad decision server. Each consumption configuration contains an end time and an optional start time that define the consumption window. Prefetch schedules automatically expire no earlier than seven days after the end time.

$sel:retrieval:CreatePrefetchSchedule', createPrefetchSchedule_retrieval - The configuration settings for retrieval of prefetched ads from the ad decision server. Only one set of prefetched ads will be retrieved and subsequently consumed for each ad break.

Request Lenses

createPrefetchSchedule_streamId :: Lens' CreatePrefetchSchedule (Maybe Text)

An optional stream identifier that MediaTailor uses to prefetch ads for multiple streams that use the same playback configuration. If StreamId is specified, MediaTailor returns all of the prefetch schedules with an exact match on StreamId. If not specified, MediaTailor returns all of the prefetch schedules for the playback configuration, regardless of StreamId.

createPrefetchSchedule_name :: Lens' CreatePrefetchSchedule Text

The identifier for the playback configuration.

createPrefetchSchedule_playbackConfigurationName :: Lens' CreatePrefetchSchedule Text

The name of the playback configuration.

createPrefetchSchedule_consumption :: Lens' CreatePrefetchSchedule PrefetchConsumption

The configuration settings for MediaTailor's consumption of the prefetched ads from the ad decision server. Each consumption configuration contains an end time and an optional start time that define the consumption window. Prefetch schedules automatically expire no earlier than seven days after the end time.

createPrefetchSchedule_retrieval :: Lens' CreatePrefetchSchedule PrefetchRetrieval

The configuration settings for retrieval of prefetched ads from the ad decision server. Only one set of prefetched ads will be retrieved and subsequently consumed for each ad break.

Destructuring the Response

data CreatePrefetchScheduleResponse

See: newCreatePrefetchScheduleResponse smart constructor.

Constructors

CreatePrefetchScheduleResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the prefetch schedule.

playbackConfigurationName :: Maybe Text

The name of the playback configuration to create the prefetch schedule for.

retrieval :: Maybe PrefetchRetrieval

A complex type that contains settings for prefetch retrieval from the ad decision server (ADS).

name :: Maybe Text

The name of the prefetch schedule. The name must be unique among all prefetch schedules that are associated with the specified playback configuration.

consumption :: Maybe PrefetchConsumption

Consumption settings determine how, and when, MediaTailor places the prefetched ads into ad breaks. Ad consumption occurs within a span of time that you define, called a consumption window. You can designate which ad breaks that MediaTailor fills with prefetch ads by setting avail matching criteria.

streamId :: Maybe Text

An optional stream identifier that you can specify in order to prefetch for multiple streams that use the same playback configuration.

httpStatus :: Int

The response's http status code.

newCreatePrefetchScheduleResponse

Arguments

:: Int
$sel:httpStatus:CreatePrefetchScheduleResponse'

-> CreatePrefetchScheduleResponse

Create a value of CreatePrefetchScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreatePrefetchScheduleResponse', createPrefetchScheduleResponse_arn - The Amazon Resource Name (ARN) of the prefetch schedule.

$sel:playbackConfigurationName:CreatePrefetchSchedule', createPrefetchScheduleResponse_playbackConfigurationName - The name of the playback configuration to create the prefetch schedule for.

$sel:retrieval:CreatePrefetchSchedule', createPrefetchScheduleResponse_retrieval - A complex type that contains settings for prefetch retrieval from the ad decision server (ADS).

$sel:name:CreatePrefetchSchedule', createPrefetchScheduleResponse_name - The name of the prefetch schedule. The name must be unique among all prefetch schedules that are associated with the specified playback configuration.

$sel:consumption:CreatePrefetchSchedule', createPrefetchScheduleResponse_consumption - Consumption settings determine how, and when, MediaTailor places the prefetched ads into ad breaks. Ad consumption occurs within a span of time that you define, called a consumption window. You can designate which ad breaks that MediaTailor fills with prefetch ads by setting avail matching criteria.

$sel:streamId:CreatePrefetchSchedule', createPrefetchScheduleResponse_streamId - An optional stream identifier that you can specify in order to prefetch for multiple streams that use the same playback configuration.

$sel:httpStatus:CreatePrefetchScheduleResponse', createPrefetchScheduleResponse_httpStatus - The response's http status code.

Response Lenses

createPrefetchScheduleResponse_arn :: Lens' CreatePrefetchScheduleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the prefetch schedule.

createPrefetchScheduleResponse_playbackConfigurationName :: Lens' CreatePrefetchScheduleResponse (Maybe Text)

The name of the playback configuration to create the prefetch schedule for.

createPrefetchScheduleResponse_retrieval :: Lens' CreatePrefetchScheduleResponse (Maybe PrefetchRetrieval)

A complex type that contains settings for prefetch retrieval from the ad decision server (ADS).

createPrefetchScheduleResponse_name :: Lens' CreatePrefetchScheduleResponse (Maybe Text)

The name of the prefetch schedule. The name must be unique among all prefetch schedules that are associated with the specified playback configuration.

createPrefetchScheduleResponse_consumption :: Lens' CreatePrefetchScheduleResponse (Maybe PrefetchConsumption)

Consumption settings determine how, and when, MediaTailor places the prefetched ads into ad breaks. Ad consumption occurs within a span of time that you define, called a consumption window. You can designate which ad breaks that MediaTailor fills with prefetch ads by setting avail matching criteria.

createPrefetchScheduleResponse_streamId :: Lens' CreatePrefetchScheduleResponse (Maybe Text)

An optional stream identifier that you can specify in order to prefetch for multiple streams that use the same playback configuration.

createPrefetchScheduleResponse_httpStatus :: Lens' CreatePrefetchScheduleResponse Int

The response's http status code.