Amazonka.MediaTailor.CreateSourceLocation

Description

Creates a source location on a specific channel.

Creating a Request

data CreateSourceLocation

See: newCreateSourceLocation smart constructor.

Constructors

CreateSourceLocation'

Fields

accessConfiguration :: Maybe AccessConfiguration

Access configuration parameters. Configures the type of authentication used to access content from your source location.

defaultSegmentDeliveryConfiguration :: Maybe DefaultSegmentDeliveryConfiguration

The optional configuration for the server that serves segments.

tags :: Maybe (HashMap Text Text)

The tags to assign to the source location.

sourceLocationName :: Text

The identifier for the source location you are working on.

httpConfiguration :: HttpConfiguration

The source's HTTP package configurations.

newCreateSourceLocation

Arguments

:: Text
$sel:sourceLocationName:CreateSourceLocation'

-> HttpConfiguration
$sel:httpConfiguration:CreateSourceLocation'

-> CreateSourceLocation

Create a value of CreateSourceLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessConfiguration:CreateSourceLocation', createSourceLocation_accessConfiguration - Access configuration parameters. Configures the type of authentication used to access content from your source location.

$sel:defaultSegmentDeliveryConfiguration:CreateSourceLocation', createSourceLocation_defaultSegmentDeliveryConfiguration - The optional configuration for the server that serves segments.

$sel:tags:CreateSourceLocation', createSourceLocation_tags - The tags to assign to the source location.

$sel:sourceLocationName:CreateSourceLocation', createSourceLocation_sourceLocationName - The identifier for the source location you are working on.

$sel:httpConfiguration:CreateSourceLocation', createSourceLocation_httpConfiguration - The source's HTTP package configurations.

Request Lenses

createSourceLocation_accessConfiguration :: Lens' CreateSourceLocation (Maybe AccessConfiguration)

Access configuration parameters. Configures the type of authentication used to access content from your source location.

createSourceLocation_defaultSegmentDeliveryConfiguration :: Lens' CreateSourceLocation (Maybe DefaultSegmentDeliveryConfiguration)

The optional configuration for the server that serves segments.

createSourceLocation_tags :: Lens' CreateSourceLocation (Maybe (HashMap Text Text))

The tags to assign to the source location.

createSourceLocation_sourceLocationName :: Lens' CreateSourceLocation Text

The identifier for the source location you are working on.

createSourceLocation_httpConfiguration :: Lens' CreateSourceLocation HttpConfiguration

The source's HTTP package configurations.

Destructuring the Response

data CreateSourceLocationResponse

See: newCreateSourceLocationResponse smart constructor.

Constructors

CreateSourceLocationResponse'

Fields

creationTime :: Maybe POSIX

The timestamp that indicates when the source location was created.

sourceLocationName :: Maybe Text

The name of the source location.

arn :: Maybe Text

The ARN of the source location.

httpConfiguration :: Maybe HttpConfiguration

The HTTP package configuration settings for the source location.

lastModifiedTime :: Maybe POSIX

The timestamp that indicates when the source location was last modified.

accessConfiguration :: Maybe AccessConfiguration

The access configuration for the source location.

defaultSegmentDeliveryConfiguration :: Maybe DefaultSegmentDeliveryConfiguration

The default segment delivery configuration settings.

tags :: Maybe (HashMap Text Text)

The tags assigned to the source location.

httpStatus :: Int

The response's http status code.

newCreateSourceLocationResponse

Arguments

:: Int
$sel:httpStatus:CreateSourceLocationResponse'

-> CreateSourceLocationResponse

Create a value of CreateSourceLocationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CreateSourceLocationResponse', createSourceLocationResponse_creationTime - The timestamp that indicates when the source location was created.

$sel:sourceLocationName:CreateSourceLocation', createSourceLocationResponse_sourceLocationName - The name of the source location.

$sel:arn:CreateSourceLocationResponse', createSourceLocationResponse_arn - The ARN of the source location.

$sel:httpConfiguration:CreateSourceLocation', createSourceLocationResponse_httpConfiguration - The HTTP package configuration settings for the source location.

$sel:lastModifiedTime:CreateSourceLocationResponse', createSourceLocationResponse_lastModifiedTime - The timestamp that indicates when the source location was last modified.

$sel:accessConfiguration:CreateSourceLocation', createSourceLocationResponse_accessConfiguration - The access configuration for the source location.

$sel:defaultSegmentDeliveryConfiguration:CreateSourceLocation', createSourceLocationResponse_defaultSegmentDeliveryConfiguration - The default segment delivery configuration settings.

$sel:tags:CreateSourceLocation', createSourceLocationResponse_tags - The tags assigned to the source location.

$sel:httpStatus:CreateSourceLocationResponse', createSourceLocationResponse_httpStatus - The response's http status code.

Response Lenses

createSourceLocationResponse_creationTime :: Lens' CreateSourceLocationResponse (Maybe UTCTime)

The timestamp that indicates when the source location was created.

createSourceLocationResponse_sourceLocationName :: Lens' CreateSourceLocationResponse (Maybe Text)

The name of the source location.

createSourceLocationResponse_arn :: Lens' CreateSourceLocationResponse (Maybe Text)

The ARN of the source location.

createSourceLocationResponse_httpConfiguration :: Lens' CreateSourceLocationResponse (Maybe HttpConfiguration)

The HTTP package configuration settings for the source location.

createSourceLocationResponse_lastModifiedTime :: Lens' CreateSourceLocationResponse (Maybe UTCTime)

The timestamp that indicates when the source location was last modified.

createSourceLocationResponse_accessConfiguration :: Lens' CreateSourceLocationResponse (Maybe AccessConfiguration)

The access configuration for the source location.

createSourceLocationResponse_defaultSegmentDeliveryConfiguration :: Lens' CreateSourceLocationResponse (Maybe DefaultSegmentDeliveryConfiguration)

The default segment delivery configuration settings.

createSourceLocationResponse_tags :: Lens' CreateSourceLocationResponse (Maybe (HashMap Text Text))

The tags assigned to the source location.

createSourceLocationResponse_httpStatus :: Lens' CreateSourceLocationResponse Int

The response's http status code.