Amazonka.MediaTailor.CreateVodSource

Description

Creates name for a specific VOD source in a source location.

Creating a Request

data CreateVodSource

See: newCreateVodSource smart constructor.

Constructors

CreateVodSource'

Fields

tags :: Maybe (HashMap Text Text)

The tags to assign to the VOD source.

sourceLocationName :: Text

The identifier for the source location you are working on.

vodSourceName :: Text

The identifier for the VOD source you are working on.

httpPackageConfigurations :: [HttpPackageConfiguration]

An array of HTTP package configuration parameters for this VOD source.

newCreateVodSource

Arguments

:: Text
$sel:sourceLocationName:CreateVodSource'

-> Text
$sel:vodSourceName:CreateVodSource'

-> CreateVodSource

Create a value of CreateVodSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateVodSource', createVodSource_tags - The tags to assign to the VOD source.

$sel:sourceLocationName:CreateVodSource', createVodSource_sourceLocationName - The identifier for the source location you are working on.

$sel:vodSourceName:CreateVodSource', createVodSource_vodSourceName - The identifier for the VOD source you are working on.

$sel:httpPackageConfigurations:CreateVodSource', createVodSource_httpPackageConfigurations - An array of HTTP package configuration parameters for this VOD source.

Request Lenses

createVodSource_tags :: Lens' CreateVodSource (Maybe (HashMap Text Text))

The tags to assign to the VOD source.

createVodSource_sourceLocationName :: Lens' CreateVodSource Text

The identifier for the source location you are working on.

createVodSource_vodSourceName :: Lens' CreateVodSource Text

The identifier for the VOD source you are working on.

createVodSource_httpPackageConfigurations :: Lens' CreateVodSource [HttpPackageConfiguration]

An array of HTTP package configuration parameters for this VOD source.

Destructuring the Response

data CreateVodSourceResponse

See: newCreateVodSourceResponse smart constructor.

Constructors

CreateVodSourceResponse'

Fields

creationTime :: Maybe POSIX

The timestamp that indicates when the VOD source was created.

sourceLocationName :: Maybe Text

The name of the source location associated with the VOD source.

arn :: Maybe Text

The ARN of the VOD source.

lastModifiedTime :: Maybe POSIX

The ARN for the VOD source.

httpPackageConfigurations :: Maybe [HttpPackageConfiguration]

The HTTP package configurations.

vodSourceName :: Maybe Text

The name of the VOD source.

tags :: Maybe (HashMap Text Text)

The tags assigned to the VOD source.

httpStatus :: Int

The response's http status code.

newCreateVodSourceResponse

Arguments

:: Int
$sel:httpStatus:CreateVodSourceResponse'

-> CreateVodSourceResponse

Create a value of CreateVodSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CreateVodSourceResponse', createVodSourceResponse_creationTime - The timestamp that indicates when the VOD source was created.

$sel:sourceLocationName:CreateVodSource', createVodSourceResponse_sourceLocationName - The name of the source location associated with the VOD source.

$sel:arn:CreateVodSourceResponse', createVodSourceResponse_arn - The ARN of the VOD source.

$sel:lastModifiedTime:CreateVodSourceResponse', createVodSourceResponse_lastModifiedTime - The ARN for the VOD source.

$sel:httpPackageConfigurations:CreateVodSource', createVodSourceResponse_httpPackageConfigurations - The HTTP package configurations.

$sel:vodSourceName:CreateVodSource', createVodSourceResponse_vodSourceName - The name of the VOD source.

$sel:tags:CreateVodSource', createVodSourceResponse_tags - The tags assigned to the VOD source.

$sel:httpStatus:CreateVodSourceResponse', createVodSourceResponse_httpStatus - The response's http status code.

Response Lenses

createVodSourceResponse_creationTime :: Lens' CreateVodSourceResponse (Maybe UTCTime)

The timestamp that indicates when the VOD source was created.

createVodSourceResponse_sourceLocationName :: Lens' CreateVodSourceResponse (Maybe Text)

The name of the source location associated with the VOD source.

createVodSourceResponse_arn :: Lens' CreateVodSourceResponse (Maybe Text)

The ARN of the VOD source.

createVodSourceResponse_lastModifiedTime :: Lens' CreateVodSourceResponse (Maybe UTCTime)

The ARN for the VOD source.

createVodSourceResponse_httpPackageConfigurations :: Lens' CreateVodSourceResponse (Maybe [HttpPackageConfiguration])

The HTTP package configurations.

createVodSourceResponse_vodSourceName :: Lens' CreateVodSourceResponse (Maybe Text)

The name of the VOD source.

createVodSourceResponse_tags :: Lens' CreateVodSourceResponse (Maybe (HashMap Text Text))

The tags assigned to the VOD source.

createVodSourceResponse_httpStatus :: Lens' CreateVodSourceResponse Int

The response's http status code.