Amazonka.MediaTailor.DescribeChannel

Description

Describes the properties of a specific channel.

Creating a Request

data DescribeChannel

See: newDescribeChannel smart constructor.

Constructors

DescribeChannel'

Fields

channelName :: Text

The identifier for the channel you are working on.

newDescribeChannel

Arguments

:: Text
$sel:channelName:DescribeChannel'

-> DescribeChannel

Create a value of DescribeChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelName:DescribeChannel', describeChannel_channelName - The identifier for the channel you are working on.

Request Lenses

describeChannel_channelName :: Lens' DescribeChannel Text

The identifier for the channel you are working on.

Destructuring the Response

data DescribeChannelResponse

See: newDescribeChannelResponse smart constructor.

Constructors

DescribeChannelResponse'

Fields

creationTime :: Maybe POSIX

The timestamp of when the channel was created.

arn :: Maybe Text

The ARN of the channel.

lastModifiedTime :: Maybe POSIX

The timestamp of when the channel was last modified.

playbackMode :: Maybe Text

The channel's playback mode.

channelName :: Maybe Text

The name of the channel.

outputs :: Maybe [ResponseOutputItem]

The channel's output properties.

channelState :: Maybe ChannelState

Indicates whether the channel is in a running state or not.

fillerSlate :: Maybe SlateSource

Contains information about the slate used to fill gaps between programs in the schedule.

tags :: Maybe (HashMap Text Text)

The tags assigned to the channel.

httpStatus :: Int

The response's http status code.

newDescribeChannelResponse

Arguments

:: Int
$sel:httpStatus:DescribeChannelResponse'

-> DescribeChannelResponse

Create a value of DescribeChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeChannelResponse', describeChannelResponse_creationTime - The timestamp of when the channel was created.

$sel:arn:DescribeChannelResponse', describeChannelResponse_arn - The ARN of the channel.

$sel:lastModifiedTime:DescribeChannelResponse', describeChannelResponse_lastModifiedTime - The timestamp of when the channel was last modified.

$sel:playbackMode:DescribeChannelResponse', describeChannelResponse_playbackMode - The channel's playback mode.

$sel:channelName:DescribeChannel', describeChannelResponse_channelName - The name of the channel.

$sel:outputs:DescribeChannelResponse', describeChannelResponse_outputs - The channel's output properties.

$sel:channelState:DescribeChannelResponse', describeChannelResponse_channelState - Indicates whether the channel is in a running state or not.

$sel:fillerSlate:DescribeChannelResponse', describeChannelResponse_fillerSlate - Contains information about the slate used to fill gaps between programs in the schedule.

$sel:tags:DescribeChannelResponse', describeChannelResponse_tags - The tags assigned to the channel.

$sel:httpStatus:DescribeChannelResponse', describeChannelResponse_httpStatus - The response's http status code.

Response Lenses

describeChannelResponse_creationTime :: Lens' DescribeChannelResponse (Maybe UTCTime)

The timestamp of when the channel was created.

describeChannelResponse_arn :: Lens' DescribeChannelResponse (Maybe Text)

The ARN of the channel.

describeChannelResponse_lastModifiedTime :: Lens' DescribeChannelResponse (Maybe UTCTime)

The timestamp of when the channel was last modified.

describeChannelResponse_playbackMode :: Lens' DescribeChannelResponse (Maybe Text)

The channel's playback mode.

describeChannelResponse_channelName :: Lens' DescribeChannelResponse (Maybe Text)

The name of the channel.

describeChannelResponse_outputs :: Lens' DescribeChannelResponse (Maybe [ResponseOutputItem])

The channel's output properties.

describeChannelResponse_channelState :: Lens' DescribeChannelResponse (Maybe ChannelState)

Indicates whether the channel is in a running state or not.

describeChannelResponse_fillerSlate :: Lens' DescribeChannelResponse (Maybe SlateSource)

Contains information about the slate used to fill gaps between programs in the schedule.

describeChannelResponse_tags :: Lens' DescribeChannelResponse (Maybe (HashMap Text Text))

The tags assigned to the channel.

describeChannelResponse_httpStatus :: Lens' DescribeChannelResponse Int

The response's http status code.