Amazonka.MediaTailor.DescribeSourceLocation

Description

Retrieves the properties of the requested source location.

Creating a Request

data DescribeSourceLocation

See: newDescribeSourceLocation smart constructor.

Constructors

DescribeSourceLocation'

Fields

sourceLocationName :: Text

The identifier for the source location you are working on.

newDescribeSourceLocation

Arguments

:: Text
$sel:sourceLocationName:DescribeSourceLocation'

-> DescribeSourceLocation

Create a value of DescribeSourceLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceLocationName:DescribeSourceLocation', describeSourceLocation_sourceLocationName - The identifier for the source location you are working on.

Request Lenses

describeSourceLocation_sourceLocationName :: Lens' DescribeSourceLocation Text

The identifier for the source location you are working on.

Destructuring the Response

data DescribeSourceLocationResponse

See: newDescribeSourceLocationResponse smart constructor.

Constructors

DescribeSourceLocationResponse'

Fields

creationTime :: Maybe POSIX

The timestamp that indicates when the source location was created.

sourceLocationName :: Maybe Text

The name of the source location.

arn :: Maybe Text

The ARN of the source location.

httpConfiguration :: Maybe HttpConfiguration

The HTTP package configuration settings for the source location.

lastModifiedTime :: Maybe POSIX

The timestamp that indicates when the source location was last modified.

accessConfiguration :: Maybe AccessConfiguration

The access configuration for the source location.

defaultSegmentDeliveryConfiguration :: Maybe DefaultSegmentDeliveryConfiguration

The default segment delivery configuration settings.

tags :: Maybe (HashMap Text Text)

The tags assigned to the source location.

httpStatus :: Int

The response's http status code.

newDescribeSourceLocationResponse

Arguments

:: Int
$sel:httpStatus:DescribeSourceLocationResponse'

-> DescribeSourceLocationResponse

Create a value of DescribeSourceLocationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeSourceLocationResponse', describeSourceLocationResponse_creationTime - The timestamp that indicates when the source location was created.

$sel:sourceLocationName:DescribeSourceLocation', describeSourceLocationResponse_sourceLocationName - The name of the source location.

$sel:arn:DescribeSourceLocationResponse', describeSourceLocationResponse_arn - The ARN of the source location.

$sel:httpConfiguration:DescribeSourceLocationResponse', describeSourceLocationResponse_httpConfiguration - The HTTP package configuration settings for the source location.

$sel:lastModifiedTime:DescribeSourceLocationResponse', describeSourceLocationResponse_lastModifiedTime - The timestamp that indicates when the source location was last modified.

$sel:accessConfiguration:DescribeSourceLocationResponse', describeSourceLocationResponse_accessConfiguration - The access configuration for the source location.

$sel:defaultSegmentDeliveryConfiguration:DescribeSourceLocationResponse', describeSourceLocationResponse_defaultSegmentDeliveryConfiguration - The default segment delivery configuration settings.

$sel:tags:DescribeSourceLocationResponse', describeSourceLocationResponse_tags - The tags assigned to the source location.

$sel:httpStatus:DescribeSourceLocationResponse', describeSourceLocationResponse_httpStatus - The response's http status code.

Response Lenses

describeSourceLocationResponse_creationTime :: Lens' DescribeSourceLocationResponse (Maybe UTCTime)

The timestamp that indicates when the source location was created.

describeSourceLocationResponse_sourceLocationName :: Lens' DescribeSourceLocationResponse (Maybe Text)

The name of the source location.

describeSourceLocationResponse_arn :: Lens' DescribeSourceLocationResponse (Maybe Text)

The ARN of the source location.

describeSourceLocationResponse_httpConfiguration :: Lens' DescribeSourceLocationResponse (Maybe HttpConfiguration)

The HTTP package configuration settings for the source location.

describeSourceLocationResponse_lastModifiedTime :: Lens' DescribeSourceLocationResponse (Maybe UTCTime)

The timestamp that indicates when the source location was last modified.

describeSourceLocationResponse_accessConfiguration :: Lens' DescribeSourceLocationResponse (Maybe AccessConfiguration)

The access configuration for the source location.

describeSourceLocationResponse_defaultSegmentDeliveryConfiguration :: Lens' DescribeSourceLocationResponse (Maybe DefaultSegmentDeliveryConfiguration)

The default segment delivery configuration settings.

describeSourceLocationResponse_tags :: Lens' DescribeSourceLocationResponse (Maybe (HashMap Text Text))

The tags assigned to the source location.

describeSourceLocationResponse_httpStatus :: Lens' DescribeSourceLocationResponse Int

The response's http status code.