Amazonka.MediaTailor.GetChannelPolicy

Description

Retrieves information about a channel's IAM policy.

Creating a Request

data GetChannelPolicy

See: newGetChannelPolicy smart constructor.

Constructors

GetChannelPolicy'

Fields

channelName :: Text

The identifier for the channel you are working on.

newGetChannelPolicy

Arguments

:: Text
$sel:channelName:GetChannelPolicy'

-> GetChannelPolicy

Create a value of GetChannelPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelName:GetChannelPolicy', getChannelPolicy_channelName - The identifier for the channel you are working on.

Request Lenses

getChannelPolicy_channelName :: Lens' GetChannelPolicy Text

The identifier for the channel you are working on.

Destructuring the Response

data GetChannelPolicyResponse

See: newGetChannelPolicyResponse smart constructor.

Constructors

GetChannelPolicyResponse'

Fields

policy :: Maybe Text

The IAM policy for the channel.

httpStatus :: Int

The response's http status code.

newGetChannelPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetChannelPolicyResponse'

-> GetChannelPolicyResponse

Create a value of GetChannelPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:GetChannelPolicyResponse', getChannelPolicyResponse_policy - The IAM policy for the channel.

$sel:httpStatus:GetChannelPolicyResponse', getChannelPolicyResponse_httpStatus - The response's http status code.

Response Lenses

getChannelPolicyResponse_policy :: Lens' GetChannelPolicyResponse (Maybe Text)

The IAM policy for the channel.

getChannelPolicyResponse_httpStatus :: Lens' GetChannelPolicyResponse Int

The response's http status code.