Amazonka.MediaTailor.GetPlaybackConfiguration

Description

Returns the playback configuration for the specified name.

Creating a Request

data GetPlaybackConfiguration

See: newGetPlaybackConfiguration smart constructor.

Constructors

GetPlaybackConfiguration'

Fields

name :: Text

The identifier for the playback configuration.

newGetPlaybackConfiguration

Arguments

:: Text
$sel:name:GetPlaybackConfiguration'

-> GetPlaybackConfiguration

Create a value of GetPlaybackConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetPlaybackConfiguration', getPlaybackConfiguration_name - The identifier for the playback configuration.

Request Lenses

getPlaybackConfiguration_name :: Lens' GetPlaybackConfiguration Text

The identifier for the playback configuration.

Destructuring the Response

data GetPlaybackConfigurationResponse

See: newGetPlaybackConfigurationResponse smart constructor.

Constructors

GetPlaybackConfigurationResponse'

Fields

playbackEndpointPrefix :: Maybe Text

The URL that the player accesses to get a manifest from AWS Elemental MediaTailor. This session will use server-side reporting.

hlsConfiguration :: Maybe HlsConfiguration

The configuration for HLS content.

personalizationThresholdSeconds :: Maybe Natural

Defines the maximum duration of underfilled ad time (in seconds) allowed in an ad break. If the duration of underfilled ad time exceeds the personalization threshold, then the personalization of the ad break is abandoned and the underlying content is shown. This feature applies to ad replacement in live and VOD streams, rather than ad insertion, because it relies on an underlying content stream. For more information about ad break behavior, including ad replacement and insertion, see Ad Behavior in AWS Elemental MediaTailor.

playbackConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) for the playback configuration.

availSuppression :: Maybe AvailSuppression

The configuration for avail suppression, also known as ad suppression. For more information about ad suppression, see Ad Suppression.

bumper :: Maybe Bumper

The configuration for bumpers. Bumpers are short audio or video clips that play at the start or before the end of an ad break. To learn more about bumpers, see Bumpers.

adDecisionServerUrl :: Maybe Text

The URL for the ad decision server (ADS). This includes the specification of static parameters and placeholders for dynamic parameters. AWS Elemental MediaTailor substitutes player-specific and session-specific parameters as needed when calling the ADS. Alternately, for testing, you can provide a static VAST URL. The maximum length is 25,000 characters.

videoContentSourceUrl :: Maybe Text

The URL prefix for the parent manifest for the stream, minus the asset ID. The maximum length is 512 characters.

dashConfiguration :: Maybe DashConfiguration

The configuration for DASH content.

manifestProcessingRules :: Maybe ManifestProcessingRules

The configuration for manifest processing rules. Manifest processing rules enable customization of the personalized manifests created by MediaTailor.

logConfiguration :: Maybe LogConfiguration

The Amazon CloudWatch log settings for a playback configuration.

livePreRollConfiguration :: Maybe LivePreRollConfiguration

The configuration for pre-roll ad insertion.

name :: Maybe Text

The identifier for the playback configuration.

sessionInitializationEndpointPrefix :: Maybe Text

The URL that the player uses to initialize a session that uses client-side reporting.

configurationAliases :: Maybe (HashMap Text (HashMap Text Text))

The player parameters and aliases used as dynamic variables during session initialization. For more information, see Domain Variables.

transcodeProfileName :: Maybe Text

The name that is used to associate this playback configuration with a custom transcode profile. This overrides the dynamic transcoding defaults of MediaTailor. Use this only if you have already set up custom profiles with the help of AWS Support.

tags :: Maybe (HashMap Text Text)

The tags assigned to the playback configuration.

slateAdUrl :: Maybe Text

The URL for a high-quality video asset to transcode and use to fill in time that's not used by ads. AWS Elemental MediaTailor shows the slate to fill in gaps in media content. Configuring the slate is optional for non-VPAID playback configurations. For VPAID, the slate is required because MediaTailor provides it in the slots designated for dynamic ad content. The slate must be a high-quality asset that contains both audio and video.

cdnConfiguration :: Maybe CdnConfiguration

The configuration for using a content delivery network (CDN), like Amazon CloudFront, for content and ad segment management.

httpStatus :: Int

The response's http status code.

newGetPlaybackConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetPlaybackConfigurationResponse'

-> GetPlaybackConfigurationResponse

Create a value of GetPlaybackConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:playbackEndpointPrefix:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_playbackEndpointPrefix - The URL that the player accesses to get a manifest from AWS Elemental MediaTailor. This session will use server-side reporting.

$sel:hlsConfiguration:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_hlsConfiguration - The configuration for HLS content.

$sel:personalizationThresholdSeconds:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_personalizationThresholdSeconds - Defines the maximum duration of underfilled ad time (in seconds) allowed in an ad break. If the duration of underfilled ad time exceeds the personalization threshold, then the personalization of the ad break is abandoned and the underlying content is shown. This feature applies to ad replacement in live and VOD streams, rather than ad insertion, because it relies on an underlying content stream. For more information about ad break behavior, including ad replacement and insertion, see Ad Behavior in AWS Elemental MediaTailor.

$sel:playbackConfigurationArn:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_playbackConfigurationArn - The Amazon Resource Name (ARN) for the playback configuration.

$sel:availSuppression:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_availSuppression - The configuration for avail suppression, also known as ad suppression. For more information about ad suppression, see Ad Suppression.

$sel:bumper:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_bumper - The configuration for bumpers. Bumpers are short audio or video clips that play at the start or before the end of an ad break. To learn more about bumpers, see Bumpers.

$sel:adDecisionServerUrl:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_adDecisionServerUrl - The URL for the ad decision server (ADS). This includes the specification of static parameters and placeholders for dynamic parameters. AWS Elemental MediaTailor substitutes player-specific and session-specific parameters as needed when calling the ADS. Alternately, for testing, you can provide a static VAST URL. The maximum length is 25,000 characters.

$sel:videoContentSourceUrl:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_videoContentSourceUrl - The URL prefix for the parent manifest for the stream, minus the asset ID. The maximum length is 512 characters.

$sel:dashConfiguration:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_dashConfiguration - The configuration for DASH content.

$sel:manifestProcessingRules:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_manifestProcessingRules - The configuration for manifest processing rules. Manifest processing rules enable customization of the personalized manifests created by MediaTailor.

$sel:logConfiguration:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_logConfiguration - The Amazon CloudWatch log settings for a playback configuration.

$sel:livePreRollConfiguration:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_livePreRollConfiguration - The configuration for pre-roll ad insertion.

$sel:name:GetPlaybackConfiguration', getPlaybackConfigurationResponse_name - The identifier for the playback configuration.

$sel:sessionInitializationEndpointPrefix:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_sessionInitializationEndpointPrefix - The URL that the player uses to initialize a session that uses client-side reporting.

$sel:configurationAliases:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_configurationAliases - The player parameters and aliases used as dynamic variables during session initialization. For more information, see Domain Variables.

$sel:transcodeProfileName:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_transcodeProfileName - The name that is used to associate this playback configuration with a custom transcode profile. This overrides the dynamic transcoding defaults of MediaTailor. Use this only if you have already set up custom profiles with the help of AWS Support.

$sel:tags:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_tags - The tags assigned to the playback configuration.

$sel:slateAdUrl:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_slateAdUrl - The URL for a high-quality video asset to transcode and use to fill in time that's not used by ads. AWS Elemental MediaTailor shows the slate to fill in gaps in media content. Configuring the slate is optional for non-VPAID playback configurations. For VPAID, the slate is required because MediaTailor provides it in the slots designated for dynamic ad content. The slate must be a high-quality asset that contains both audio and video.

$sel:cdnConfiguration:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_cdnConfiguration - The configuration for using a content delivery network (CDN), like Amazon CloudFront, for content and ad segment management.

$sel:httpStatus:GetPlaybackConfigurationResponse', getPlaybackConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getPlaybackConfigurationResponse_playbackEndpointPrefix :: Lens' GetPlaybackConfigurationResponse (Maybe Text)

The URL that the player accesses to get a manifest from AWS Elemental MediaTailor. This session will use server-side reporting.

getPlaybackConfigurationResponse_hlsConfiguration :: Lens' GetPlaybackConfigurationResponse (Maybe HlsConfiguration)

The configuration for HLS content.

getPlaybackConfigurationResponse_personalizationThresholdSeconds :: Lens' GetPlaybackConfigurationResponse (Maybe Natural)

Defines the maximum duration of underfilled ad time (in seconds) allowed in an ad break. If the duration of underfilled ad time exceeds the personalization threshold, then the personalization of the ad break is abandoned and the underlying content is shown. This feature applies to ad replacement in live and VOD streams, rather than ad insertion, because it relies on an underlying content stream. For more information about ad break behavior, including ad replacement and insertion, see Ad Behavior in AWS Elemental MediaTailor.

getPlaybackConfigurationResponse_playbackConfigurationArn :: Lens' GetPlaybackConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) for the playback configuration.

getPlaybackConfigurationResponse_availSuppression :: Lens' GetPlaybackConfigurationResponse (Maybe AvailSuppression)

The configuration for avail suppression, also known as ad suppression. For more information about ad suppression, see Ad Suppression.

getPlaybackConfigurationResponse_bumper :: Lens' GetPlaybackConfigurationResponse (Maybe Bumper)

The configuration for bumpers. Bumpers are short audio or video clips that play at the start or before the end of an ad break. To learn more about bumpers, see Bumpers.

getPlaybackConfigurationResponse_adDecisionServerUrl :: Lens' GetPlaybackConfigurationResponse (Maybe Text)

The URL for the ad decision server (ADS). This includes the specification of static parameters and placeholders for dynamic parameters. AWS Elemental MediaTailor substitutes player-specific and session-specific parameters as needed when calling the ADS. Alternately, for testing, you can provide a static VAST URL. The maximum length is 25,000 characters.

getPlaybackConfigurationResponse_videoContentSourceUrl :: Lens' GetPlaybackConfigurationResponse (Maybe Text)

The URL prefix for the parent manifest for the stream, minus the asset ID. The maximum length is 512 characters.

getPlaybackConfigurationResponse_dashConfiguration :: Lens' GetPlaybackConfigurationResponse (Maybe DashConfiguration)

The configuration for DASH content.

getPlaybackConfigurationResponse_manifestProcessingRules :: Lens' GetPlaybackConfigurationResponse (Maybe ManifestProcessingRules)

The configuration for manifest processing rules. Manifest processing rules enable customization of the personalized manifests created by MediaTailor.

getPlaybackConfigurationResponse_logConfiguration :: Lens' GetPlaybackConfigurationResponse (Maybe LogConfiguration)

The Amazon CloudWatch log settings for a playback configuration.

getPlaybackConfigurationResponse_livePreRollConfiguration :: Lens' GetPlaybackConfigurationResponse (Maybe LivePreRollConfiguration)

The configuration for pre-roll ad insertion.

getPlaybackConfigurationResponse_name :: Lens' GetPlaybackConfigurationResponse (Maybe Text)

The identifier for the playback configuration.

getPlaybackConfigurationResponse_sessionInitializationEndpointPrefix :: Lens' GetPlaybackConfigurationResponse (Maybe Text)

The URL that the player uses to initialize a session that uses client-side reporting.

getPlaybackConfigurationResponse_configurationAliases :: Lens' GetPlaybackConfigurationResponse (Maybe (HashMap Text (HashMap Text Text)))

The player parameters and aliases used as dynamic variables during session initialization. For more information, see Domain Variables.

getPlaybackConfigurationResponse_transcodeProfileName :: Lens' GetPlaybackConfigurationResponse (Maybe Text)

The name that is used to associate this playback configuration with a custom transcode profile. This overrides the dynamic transcoding defaults of MediaTailor. Use this only if you have already set up custom profiles with the help of AWS Support.

getPlaybackConfigurationResponse_tags :: Lens' GetPlaybackConfigurationResponse (Maybe (HashMap Text Text))

The tags assigned to the playback configuration.

getPlaybackConfigurationResponse_slateAdUrl :: Lens' GetPlaybackConfigurationResponse (Maybe Text)

The URL for a high-quality video asset to transcode and use to fill in time that's not used by ads. AWS Elemental MediaTailor shows the slate to fill in gaps in media content. Configuring the slate is optional for non-VPAID playback configurations. For VPAID, the slate is required because MediaTailor provides it in the slots designated for dynamic ad content. The slate must be a high-quality asset that contains both audio and video.

getPlaybackConfigurationResponse_cdnConfiguration :: Lens' GetPlaybackConfigurationResponse (Maybe CdnConfiguration)

The configuration for using a content delivery network (CDN), like Amazon CloudFront, for content and ad segment management.

getPlaybackConfigurationResponse_httpStatus :: Lens' GetPlaybackConfigurationResponse Int

The response's http status code.