Amazonka.MediaTailor.ListPrefetchSchedules

Description

Creates a new prefetch schedule.

This operation returns paginated results.

Creating a Request

data ListPrefetchSchedules

See: newListPrefetchSchedules smart constructor.

Constructors

ListPrefetchSchedules'

Fields

nextToken :: Maybe Text

(Optional) If the playback configuration has more than MaxResults prefetch schedules, use NextToken to get the second and subsequent pages of results.

For the first ListPrefetchSchedulesRequest request, omit this value.

For the second and subsequent requests, get the value of NextToken from the previous response and specify that value for NextToken in the request.

If the previous response didn't include a NextToken element, there are no more prefetch schedules to get.

maxResults :: Maybe Natural

The maximum number of prefetch schedules that you want MediaTailor to return in response to the current request. If the playback configuration has more than MaxResults prefetch schedules, use the value of NextToken in the response to get the next page of results.

streamId :: Maybe Text

An optional filtering parameter whereby MediaTailor filters the prefetch schedules to include only specific streams.

playbackConfigurationName :: Text

The name of the playback configuration.

newListPrefetchSchedules

Arguments

:: Text
$sel:playbackConfigurationName:ListPrefetchSchedules'

-> ListPrefetchSchedules

Create a value of ListPrefetchSchedules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPrefetchSchedules', listPrefetchSchedules_nextToken - (Optional) If the playback configuration has more than MaxResults prefetch schedules, use NextToken to get the second and subsequent pages of results.

For the first ListPrefetchSchedulesRequest request, omit this value.

For the second and subsequent requests, get the value of NextToken from the previous response and specify that value for NextToken in the request.

If the previous response didn't include a NextToken element, there are no more prefetch schedules to get.

$sel:maxResults:ListPrefetchSchedules', listPrefetchSchedules_maxResults - The maximum number of prefetch schedules that you want MediaTailor to return in response to the current request. If the playback configuration has more than MaxResults prefetch schedules, use the value of NextToken in the response to get the next page of results.

$sel:streamId:ListPrefetchSchedules', listPrefetchSchedules_streamId - An optional filtering parameter whereby MediaTailor filters the prefetch schedules to include only specific streams.

$sel:playbackConfigurationName:ListPrefetchSchedules', listPrefetchSchedules_playbackConfigurationName - The name of the playback configuration.

Request Lenses

listPrefetchSchedules_nextToken :: Lens' ListPrefetchSchedules (Maybe Text)

(Optional) If the playback configuration has more than MaxResults prefetch schedules, use NextToken to get the second and subsequent pages of results.

For the first ListPrefetchSchedulesRequest request, omit this value.

For the second and subsequent requests, get the value of NextToken from the previous response and specify that value for NextToken in the request.

If the previous response didn't include a NextToken element, there are no more prefetch schedules to get.

listPrefetchSchedules_maxResults :: Lens' ListPrefetchSchedules (Maybe Natural)

The maximum number of prefetch schedules that you want MediaTailor to return in response to the current request. If the playback configuration has more than MaxResults prefetch schedules, use the value of NextToken in the response to get the next page of results.

listPrefetchSchedules_streamId :: Lens' ListPrefetchSchedules (Maybe Text)

An optional filtering parameter whereby MediaTailor filters the prefetch schedules to include only specific streams.

listPrefetchSchedules_playbackConfigurationName :: Lens' ListPrefetchSchedules Text

The name of the playback configuration.

Destructuring the Response

data ListPrefetchSchedulesResponse

See: newListPrefetchSchedulesResponse smart constructor.

Constructors

ListPrefetchSchedulesResponse'

Fields

items :: Maybe [PrefetchSchedule]

Lists the prefetch schedules. An empty Items list doesn't mean there aren't more items to fetch, just that that page was empty.

nextToken :: Maybe Text

The value that you will use forNextToken in the next ListPrefetchSchedulesRequest request.

httpStatus :: Int

The response's http status code.

newListPrefetchSchedulesResponse

Arguments

:: Int
$sel:httpStatus:ListPrefetchSchedulesResponse'

-> ListPrefetchSchedulesResponse

Create a value of ListPrefetchSchedulesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:ListPrefetchSchedulesResponse', listPrefetchSchedulesResponse_items - Lists the prefetch schedules. An empty Items list doesn't mean there aren't more items to fetch, just that that page was empty.

$sel:nextToken:ListPrefetchSchedules', listPrefetchSchedulesResponse_nextToken - The value that you will use forNextToken in the next ListPrefetchSchedulesRequest request.

$sel:httpStatus:ListPrefetchSchedulesResponse', listPrefetchSchedulesResponse_httpStatus - The response's http status code.

Response Lenses

listPrefetchSchedulesResponse_items :: Lens' ListPrefetchSchedulesResponse (Maybe [PrefetchSchedule])

Lists the prefetch schedules. An empty Items list doesn't mean there aren't more items to fetch, just that that page was empty.

listPrefetchSchedulesResponse_nextToken :: Lens' ListPrefetchSchedulesResponse (Maybe Text)

The value that you will use forNextToken in the next ListPrefetchSchedulesRequest request.

listPrefetchSchedulesResponse_httpStatus :: Lens' ListPrefetchSchedulesResponse Int

The response's http status code.