Amazonka.MediaTailor.StopChannel

Description

Stops a specific channel.

Creating a Request

data StopChannel

See: newStopChannel smart constructor.

Constructors

StopChannel'

Fields

channelName :: Text

The identifier for the channel you are working on.

newStopChannel

Arguments

:: Text
$sel:channelName:StopChannel'

-> StopChannel

Create a value of StopChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelName:StopChannel', stopChannel_channelName - The identifier for the channel you are working on.

Request Lenses

stopChannel_channelName :: Lens' StopChannel Text

The identifier for the channel you are working on.

Destructuring the Response

data StopChannelResponse

See: newStopChannelResponse smart constructor.

Constructors

StopChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

newStopChannelResponse

Arguments

:: Int
$sel:httpStatus:StopChannelResponse'

-> StopChannelResponse

Create a value of StopChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StopChannelResponse', stopChannelResponse_httpStatus - The response's http status code.

Response Lenses

stopChannelResponse_httpStatus :: Lens' StopChannelResponse Int

The response's http status code.