Amazonka.MediaTailor.Types.Alert

Description

Documentation

data Alert

Alert configuration parameters.

See: newAlert smart constructor.

Constructors

Alert'

Fields

resourceArn :: Text

The Amazon Resource Name (ARN) of the resource.

alertCode :: Text

The code for the alert. For example, NOT_PROCESSED.

lastModifiedTime :: POSIX

The timestamp when the alert was last modified.

relatedResourceArns :: [Text]

The Amazon Resource Names (ARNs) related to this alert.

alertMessage :: Text

If an alert is generated for a resource, an explanation of the reason for the alert.

newAlert

Arguments

:: Text
$sel:resourceArn:Alert'

-> Text
$sel:alertCode:Alert'

-> UTCTime
$sel:lastModifiedTime:Alert'

-> Text
$sel:alertMessage:Alert'

-> Alert

Create a value of Alert with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:Alert', alert_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:alertCode:Alert', alert_alertCode - The code for the alert. For example, NOT_PROCESSED.

$sel:lastModifiedTime:Alert', alert_lastModifiedTime - The timestamp when the alert was last modified.

$sel:relatedResourceArns:Alert', alert_relatedResourceArns - The Amazon Resource Names (ARNs) related to this alert.

$sel:alertMessage:Alert', alert_alertMessage - If an alert is generated for a resource, an explanation of the reason for the alert.

alert_resourceArn :: Lens' Alert Text

The Amazon Resource Name (ARN) of the resource.

alert_alertCode :: Lens' Alert Text

The code for the alert. For example, NOT_PROCESSED.

alert_lastModifiedTime :: Lens' Alert UTCTime

The timestamp when the alert was last modified.

alert_relatedResourceArns :: Lens' Alert [Text]

The Amazon Resource Names (ARNs) related to this alert.

alert_alertMessage :: Lens' Alert Text

If an alert is generated for a resource, an explanation of the reason for the alert.