Amazonka.MediaTailor.Types.DashPlaylistSettings

Description

Documentation

data DashPlaylistSettings

Dash manifest configuration parameters.

See: newDashPlaylistSettings smart constructor.

Constructors

DashPlaylistSettings'

Fields

minBufferTimeSeconds :: Maybe Int

Minimum amount of content (measured in seconds) that a player must keep available in the buffer. Minimum value: 2 seconds. Maximum value: 60 seconds.

minUpdatePeriodSeconds :: Maybe Int

Minimum amount of time (in seconds) that the player should wait before requesting updates to the manifest. Minimum value: 2 seconds. Maximum value: 60 seconds.

suggestedPresentationDelaySeconds :: Maybe Int

Amount of time (in seconds) that the player should be from the live point at the end of the manifest. Minimum value: 2 seconds. Maximum value: 60 seconds.

manifestWindowSeconds :: Maybe Int

The total duration (in seconds) of each manifest. Minimum value: 30 seconds. Maximum value: 3600 seconds.

newDashPlaylistSettings :: DashPlaylistSettings

Create a value of DashPlaylistSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minBufferTimeSeconds:DashPlaylistSettings', dashPlaylistSettings_minBufferTimeSeconds - Minimum amount of content (measured in seconds) that a player must keep available in the buffer. Minimum value: 2 seconds. Maximum value: 60 seconds.

$sel:minUpdatePeriodSeconds:DashPlaylistSettings', dashPlaylistSettings_minUpdatePeriodSeconds - Minimum amount of time (in seconds) that the player should wait before requesting updates to the manifest. Minimum value: 2 seconds. Maximum value: 60 seconds.

$sel:suggestedPresentationDelaySeconds:DashPlaylistSettings', dashPlaylistSettings_suggestedPresentationDelaySeconds - Amount of time (in seconds) that the player should be from the live point at the end of the manifest. Minimum value: 2 seconds. Maximum value: 60 seconds.

$sel:manifestWindowSeconds:DashPlaylistSettings', dashPlaylistSettings_manifestWindowSeconds - The total duration (in seconds) of each manifest. Minimum value: 30 seconds. Maximum value: 3600 seconds.

dashPlaylistSettings_minBufferTimeSeconds :: Lens' DashPlaylistSettings (Maybe Int)

Minimum amount of content (measured in seconds) that a player must keep available in the buffer. Minimum value: 2 seconds. Maximum value: 60 seconds.

dashPlaylistSettings_minUpdatePeriodSeconds :: Lens' DashPlaylistSettings (Maybe Int)

Minimum amount of time (in seconds) that the player should wait before requesting updates to the manifest. Minimum value: 2 seconds. Maximum value: 60 seconds.

dashPlaylistSettings_suggestedPresentationDelaySeconds :: Lens' DashPlaylistSettings (Maybe Int)

Amount of time (in seconds) that the player should be from the live point at the end of the manifest. Minimum value: 2 seconds. Maximum value: 60 seconds.

dashPlaylistSettings_manifestWindowSeconds :: Lens' DashPlaylistSettings (Maybe Int)

The total duration (in seconds) of each manifest. Minimum value: 30 seconds. Maximum value: 3600 seconds.