Amazonka.MediaTailor.Types.ResponseOutputItem

Description

Documentation

data ResponseOutputItem

This response includes only the "property" : "type" property.

See: newResponseOutputItem smart constructor.

Constructors

ResponseOutputItem'

Fields

hlsPlaylistSettings :: Maybe HlsPlaylistSettings

HLS manifest configuration settings.

dashPlaylistSettings :: Maybe DashPlaylistSettings

DASH manifest configuration settings.

manifestName :: Text

The name of the manifest for the channel that will appear in the channel output's playback URL.

playbackUrl :: Text

The URL used for playback by content players.

sourceGroup :: Text

A string used to associate a package configuration source group with a channel output.

newResponseOutputItem

Arguments

:: Text
$sel:manifestName:ResponseOutputItem'

-> Text
$sel:playbackUrl:ResponseOutputItem'

-> Text
$sel:sourceGroup:ResponseOutputItem'

-> ResponseOutputItem

Create a value of ResponseOutputItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hlsPlaylistSettings:ResponseOutputItem', responseOutputItem_hlsPlaylistSettings - HLS manifest configuration settings.

$sel:dashPlaylistSettings:ResponseOutputItem', responseOutputItem_dashPlaylistSettings - DASH manifest configuration settings.

$sel:manifestName:ResponseOutputItem', responseOutputItem_manifestName - The name of the manifest for the channel that will appear in the channel output's playback URL.

$sel:playbackUrl:ResponseOutputItem', responseOutputItem_playbackUrl - The URL used for playback by content players.

$sel:sourceGroup:ResponseOutputItem', responseOutputItem_sourceGroup - A string used to associate a package configuration source group with a channel output.

responseOutputItem_hlsPlaylistSettings :: Lens' ResponseOutputItem (Maybe HlsPlaylistSettings)

HLS manifest configuration settings.

responseOutputItem_dashPlaylistSettings :: Lens' ResponseOutputItem (Maybe DashPlaylistSettings)

DASH manifest configuration settings.

responseOutputItem_manifestName :: Lens' ResponseOutputItem Text

The name of the manifest for the channel that will appear in the channel output's playback URL.

responseOutputItem_playbackUrl :: Lens' ResponseOutputItem Text

The URL used for playback by content players.

responseOutputItem_sourceGroup :: Lens' ResponseOutputItem Text

A string used to associate a package configuration source group with a channel output.