Amazonka.MediaTailor.Types.SourceLocation

Description

Documentation

data SourceLocation

This response includes only the "type" : "object" property.

See: newSourceLocation smart constructor.

Constructors

SourceLocation'

Fields

creationTime :: Maybe POSIX

The timestamp that indicates when the source location was created.

lastModifiedTime :: Maybe POSIX

The timestamp that indicates when the source location was last modified.

accessConfiguration :: Maybe AccessConfiguration

The access configuration for the source location.

defaultSegmentDeliveryConfiguration :: Maybe DefaultSegmentDeliveryConfiguration

The default segment delivery configuration.

tags :: Maybe (HashMap Text Text)

The tags assigned to the source location.

sourceLocationName :: Text

The name of the source location.

httpConfiguration :: HttpConfiguration

The HTTP configuration for the source location.

arn :: Text

The ARN of the SourceLocation.

newSourceLocation

Arguments

:: Text
$sel:sourceLocationName:SourceLocation'

-> HttpConfiguration
$sel:httpConfiguration:SourceLocation'

-> Text
$sel:arn:SourceLocation'

-> SourceLocation

Create a value of SourceLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:SourceLocation', sourceLocation_creationTime - The timestamp that indicates when the source location was created.

$sel:lastModifiedTime:SourceLocation', sourceLocation_lastModifiedTime - The timestamp that indicates when the source location was last modified.

$sel:accessConfiguration:SourceLocation', sourceLocation_accessConfiguration - The access configuration for the source location.

$sel:defaultSegmentDeliveryConfiguration:SourceLocation', sourceLocation_defaultSegmentDeliveryConfiguration - The default segment delivery configuration.

$sel:tags:SourceLocation', sourceLocation_tags - The tags assigned to the source location.

$sel:sourceLocationName:SourceLocation', sourceLocation_sourceLocationName - The name of the source location.

$sel:httpConfiguration:SourceLocation', sourceLocation_httpConfiguration - The HTTP configuration for the source location.

$sel:arn:SourceLocation', sourceLocation_arn - The ARN of the SourceLocation.

sourceLocation_creationTime :: Lens' SourceLocation (Maybe UTCTime)

The timestamp that indicates when the source location was created.

sourceLocation_lastModifiedTime :: Lens' SourceLocation (Maybe UTCTime)

The timestamp that indicates when the source location was last modified.

sourceLocation_accessConfiguration :: Lens' SourceLocation (Maybe AccessConfiguration)

The access configuration for the source location.

sourceLocation_defaultSegmentDeliveryConfiguration :: Lens' SourceLocation (Maybe DefaultSegmentDeliveryConfiguration)

The default segment delivery configuration.

sourceLocation_tags :: Lens' SourceLocation (Maybe (HashMap Text Text))

The tags assigned to the source location.

sourceLocation_sourceLocationName :: Lens' SourceLocation Text

The name of the source location.

sourceLocation_httpConfiguration :: Lens' SourceLocation HttpConfiguration

The HTTP configuration for the source location.

sourceLocation_arn :: Lens' SourceLocation Text

The ARN of the SourceLocation.