Amazonka.MemoryDb.CreateCluster

Description

Creates a cluster. All nodes in the cluster run the same protocol-compliant engine software.

Creating a Request

data CreateCluster

See: newCreateCluster smart constructor.

Constructors

CreateCluster'

Fields

engineVersion :: Maybe Text

The version number of the Redis engine to be used for the cluster.

securityGroupIds :: Maybe [Text]

A list of security group names to associate with this cluster.

snapshotArns :: Maybe [Text]

A list of Amazon Resource Names (ARN) that uniquely identify the RDB snapshot files stored in Amazon S3. The snapshot files are used to populate the new cluster. The Amazon S3 object name in the ARN cannot contain any commas.

autoMinorVersionUpgrade :: Maybe Bool

When set to true, the cluster will automatically receive minor engine version upgrades after launch.

snsTopicArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications are sent.

snapshotWindow :: Maybe Text

The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your shard.

Example: 05:00-09:00

If you do not specify this parameter, MemoryDB automatically chooses an appropriate time range.

subnetGroupName :: Maybe Text

The name of the subnet group to be used for the cluster.

tLSEnabled :: Maybe Bool

A flag to enable in-transit encryption on the cluster.

numShards :: Maybe Int

The number of shards the cluster will contain. The default value is 1.

maintenanceWindow :: Maybe Text

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

kmsKeyId :: Maybe Text

The ID of the KMS key used to encrypt the cluster.

snapshotRetentionLimit :: Maybe Int

The number of days for which MemoryDB retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

snapshotName :: Maybe Text

The name of a snapshot from which to restore data into the new cluster. The snapshot status changes to restoring while the new cluster is being created.

description :: Maybe Text

An optional description of the cluster.

numReplicasPerShard :: Maybe Int

The number of replicas to apply to each shard. The default value is 1. The maximum is 5.

tags :: Maybe [Tag]

A list of tags to be added to this resource. Tags are comma-separated key,value pairs (e.g. Key=myKey, Value=myKeyValue. You can include multiple tags as shown following: Key=myKey, Value=myKeyValue Key=mySecondKey, Value=mySecondKeyValue.

port :: Maybe Int

The port number on which each of the nodes accepts connections.

parameterGroupName :: Maybe Text

The name of the parameter group associated with the cluster.

clusterName :: Text

The name of the cluster. This value must be unique as it also serves as the cluster identifier.

nodeType :: Text

The compute and memory capacity of the nodes in the cluster.

aCLName :: Text

The name of the Access Control List to associate with the cluster.

newCreateCluster

Arguments

:: Text
$sel:clusterName:CreateCluster'

-> Text
$sel:nodeType:CreateCluster'

-> Text
$sel:aCLName:CreateCluster'

-> CreateCluster

Create a value of CreateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:CreateCluster', createCluster_engineVersion - The version number of the Redis engine to be used for the cluster.

$sel:securityGroupIds:CreateCluster', createCluster_securityGroupIds - A list of security group names to associate with this cluster.

$sel:snapshotArns:CreateCluster', createCluster_snapshotArns - A list of Amazon Resource Names (ARN) that uniquely identify the RDB snapshot files stored in Amazon S3. The snapshot files are used to populate the new cluster. The Amazon S3 object name in the ARN cannot contain any commas.

$sel:autoMinorVersionUpgrade:CreateCluster', createCluster_autoMinorVersionUpgrade - When set to true, the cluster will automatically receive minor engine version upgrades after launch.

$sel:snsTopicArn:CreateCluster', createCluster_snsTopicArn - The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications are sent.

$sel:snapshotWindow:CreateCluster', createCluster_snapshotWindow - The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your shard.

Example: 05:00-09:00

If you do not specify this parameter, MemoryDB automatically chooses an appropriate time range.

$sel:subnetGroupName:CreateCluster', createCluster_subnetGroupName - The name of the subnet group to be used for the cluster.

$sel:tLSEnabled:CreateCluster', createCluster_tLSEnabled - A flag to enable in-transit encryption on the cluster.

$sel:numShards:CreateCluster', createCluster_numShards - The number of shards the cluster will contain. The default value is 1.

$sel:maintenanceWindow:CreateCluster', createCluster_maintenanceWindow - Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

$sel:kmsKeyId:CreateCluster', createCluster_kmsKeyId - The ID of the KMS key used to encrypt the cluster.

$sel:snapshotRetentionLimit:CreateCluster', createCluster_snapshotRetentionLimit - The number of days for which MemoryDB retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

$sel:snapshotName:CreateCluster', createCluster_snapshotName - The name of a snapshot from which to restore data into the new cluster. The snapshot status changes to restoring while the new cluster is being created.

$sel:description:CreateCluster', createCluster_description - An optional description of the cluster.

$sel:numReplicasPerShard:CreateCluster', createCluster_numReplicasPerShard - The number of replicas to apply to each shard. The default value is 1. The maximum is 5.

$sel:tags:CreateCluster', createCluster_tags - A list of tags to be added to this resource. Tags are comma-separated key,value pairs (e.g. Key=myKey, Value=myKeyValue. You can include multiple tags as shown following: Key=myKey, Value=myKeyValue Key=mySecondKey, Value=mySecondKeyValue.

$sel:port:CreateCluster', createCluster_port - The port number on which each of the nodes accepts connections.

$sel:parameterGroupName:CreateCluster', createCluster_parameterGroupName - The name of the parameter group associated with the cluster.

$sel:clusterName:CreateCluster', createCluster_clusterName - The name of the cluster. This value must be unique as it also serves as the cluster identifier.

$sel:nodeType:CreateCluster', createCluster_nodeType - The compute and memory capacity of the nodes in the cluster.

$sel:aCLName:CreateCluster', createCluster_aCLName - The name of the Access Control List to associate with the cluster.

Request Lenses

createCluster_engineVersion :: Lens' CreateCluster (Maybe Text)

The version number of the Redis engine to be used for the cluster.

createCluster_securityGroupIds :: Lens' CreateCluster (Maybe [Text])

A list of security group names to associate with this cluster.

createCluster_snapshotArns :: Lens' CreateCluster (Maybe [Text])

A list of Amazon Resource Names (ARN) that uniquely identify the RDB snapshot files stored in Amazon S3. The snapshot files are used to populate the new cluster. The Amazon S3 object name in the ARN cannot contain any commas.

createCluster_autoMinorVersionUpgrade :: Lens' CreateCluster (Maybe Bool)

When set to true, the cluster will automatically receive minor engine version upgrades after launch.

createCluster_snsTopicArn :: Lens' CreateCluster (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications are sent.

createCluster_snapshotWindow :: Lens' CreateCluster (Maybe Text)

The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your shard.

Example: 05:00-09:00

If you do not specify this parameter, MemoryDB automatically chooses an appropriate time range.

createCluster_subnetGroupName :: Lens' CreateCluster (Maybe Text)

The name of the subnet group to be used for the cluster.

createCluster_tLSEnabled :: Lens' CreateCluster (Maybe Bool)

A flag to enable in-transit encryption on the cluster.

createCluster_numShards :: Lens' CreateCluster (Maybe Int)

The number of shards the cluster will contain. The default value is 1.

createCluster_maintenanceWindow :: Lens' CreateCluster (Maybe Text)

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

createCluster_kmsKeyId :: Lens' CreateCluster (Maybe Text)

The ID of the KMS key used to encrypt the cluster.

createCluster_snapshotRetentionLimit :: Lens' CreateCluster (Maybe Int)

The number of days for which MemoryDB retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

createCluster_snapshotName :: Lens' CreateCluster (Maybe Text)

The name of a snapshot from which to restore data into the new cluster. The snapshot status changes to restoring while the new cluster is being created.

createCluster_description :: Lens' CreateCluster (Maybe Text)

An optional description of the cluster.

createCluster_numReplicasPerShard :: Lens' CreateCluster (Maybe Int)

The number of replicas to apply to each shard. The default value is 1. The maximum is 5.

createCluster_tags :: Lens' CreateCluster (Maybe [Tag])

A list of tags to be added to this resource. Tags are comma-separated key,value pairs (e.g. Key=myKey, Value=myKeyValue. You can include multiple tags as shown following: Key=myKey, Value=myKeyValue Key=mySecondKey, Value=mySecondKeyValue.

createCluster_port :: Lens' CreateCluster (Maybe Int)

The port number on which each of the nodes accepts connections.

createCluster_parameterGroupName :: Lens' CreateCluster (Maybe Text)

The name of the parameter group associated with the cluster.

createCluster_clusterName :: Lens' CreateCluster Text

The name of the cluster. This value must be unique as it also serves as the cluster identifier.

createCluster_nodeType :: Lens' CreateCluster Text

The compute and memory capacity of the nodes in the cluster.

createCluster_aCLName :: Lens' CreateCluster Text

The name of the Access Control List to associate with the cluster.

Destructuring the Response

data CreateClusterResponse

See: newCreateClusterResponse smart constructor.

Constructors

CreateClusterResponse'

Fields

cluster :: Maybe Cluster

The newly-created cluster.

httpStatus :: Int

The response's http status code.

newCreateClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateClusterResponse'

-> CreateClusterResponse

Create a value of CreateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:CreateClusterResponse', createClusterResponse_cluster - The newly-created cluster.

$sel:httpStatus:CreateClusterResponse', createClusterResponse_httpStatus - The response's http status code.

Response Lenses

createClusterResponse_cluster :: Lens' CreateClusterResponse (Maybe Cluster)

The newly-created cluster.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.