Amazonka.MemoryDb.CreateSnapshot

Description

Creates a copy of an entire cluster at a specific moment in time.

Creating a Request

data CreateSnapshot

See: newCreateSnapshot smart constructor.

Constructors

CreateSnapshot'

Fields

kmsKeyId :: Maybe Text

The ID of the KMS key used to encrypt the snapshot.

tags :: Maybe [Tag]

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

clusterName :: Text

The snapshot is created from this cluster.

snapshotName :: Text

A name for the snapshot being created.

newCreateSnapshot

Arguments

:: Text
$sel:clusterName:CreateSnapshot'

-> Text
$sel:snapshotName:CreateSnapshot'

-> CreateSnapshot

Create a value of CreateSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:CreateSnapshot', createSnapshot_kmsKeyId - The ID of the KMS key used to encrypt the snapshot.

$sel:tags:CreateSnapshot', createSnapshot_tags - A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

$sel:clusterName:CreateSnapshot', createSnapshot_clusterName - The snapshot is created from this cluster.

$sel:snapshotName:CreateSnapshot', createSnapshot_snapshotName - A name for the snapshot being created.

Request Lenses

createSnapshot_kmsKeyId :: Lens' CreateSnapshot (Maybe Text)

The ID of the KMS key used to encrypt the snapshot.

createSnapshot_tags :: Lens' CreateSnapshot (Maybe [Tag])

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

createSnapshot_clusterName :: Lens' CreateSnapshot Text

The snapshot is created from this cluster.

createSnapshot_snapshotName :: Lens' CreateSnapshot Text

A name for the snapshot being created.

Destructuring the Response

data CreateSnapshotResponse

See: newCreateSnapshotResponse smart constructor.

Constructors

CreateSnapshotResponse'

Fields

snapshot :: Maybe Snapshot

The newly-created snapshot.

httpStatus :: Int

The response's http status code.

newCreateSnapshotResponse

Arguments

:: Int
$sel:httpStatus:CreateSnapshotResponse'

-> CreateSnapshotResponse

Create a value of CreateSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshot:CreateSnapshotResponse', createSnapshotResponse_snapshot - The newly-created snapshot.

$sel:httpStatus:CreateSnapshotResponse', createSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

createSnapshotResponse_snapshot :: Lens' CreateSnapshotResponse (Maybe Snapshot)

The newly-created snapshot.

createSnapshotResponse_httpStatus :: Lens' CreateSnapshotResponse Int

The response's http status code.