Amazonka.MemoryDb.DescribeServiceUpdates

Description

Returns details of the service updates

Creating a Request

data DescribeServiceUpdates

See: newDescribeServiceUpdates smart constructor.

Constructors

DescribeServiceUpdates'

Fields

status :: Maybe [ServiceUpdateStatus]

The status(es) of the service updates to filter on

serviceUpdateName :: Maybe Text

The unique ID of the service update to describe.

clusterNames :: Maybe [Text]

The list of cluster names to identify service updates to apply

nextToken :: Maybe Text

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

maxResults :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

newDescribeServiceUpdates :: DescribeServiceUpdates

Create a value of DescribeServiceUpdates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeServiceUpdates', describeServiceUpdates_status - The status(es) of the service updates to filter on

$sel:serviceUpdateName:DescribeServiceUpdates', describeServiceUpdates_serviceUpdateName - The unique ID of the service update to describe.

$sel:clusterNames:DescribeServiceUpdates', describeServiceUpdates_clusterNames - The list of cluster names to identify service updates to apply

$sel:nextToken:DescribeServiceUpdates', describeServiceUpdates_nextToken - An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

$sel:maxResults:DescribeServiceUpdates', describeServiceUpdates_maxResults - The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

Request Lenses

describeServiceUpdates_status :: Lens' DescribeServiceUpdates (Maybe [ServiceUpdateStatus])

The status(es) of the service updates to filter on

describeServiceUpdates_serviceUpdateName :: Lens' DescribeServiceUpdates (Maybe Text)

The unique ID of the service update to describe.

describeServiceUpdates_clusterNames :: Lens' DescribeServiceUpdates (Maybe [Text])

The list of cluster names to identify service updates to apply

describeServiceUpdates_nextToken :: Lens' DescribeServiceUpdates (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeServiceUpdates_maxResults :: Lens' DescribeServiceUpdates (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

Destructuring the Response

data DescribeServiceUpdatesResponse

See: newDescribeServiceUpdatesResponse smart constructor.

Constructors

DescribeServiceUpdatesResponse'

Fields

serviceUpdates :: Maybe [ServiceUpdate]

A list of service updates

nextToken :: Maybe Text

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

httpStatus :: Int

The response's http status code.

newDescribeServiceUpdatesResponse

Arguments

:: Int
$sel:httpStatus:DescribeServiceUpdatesResponse'

-> DescribeServiceUpdatesResponse

Create a value of DescribeServiceUpdatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceUpdates:DescribeServiceUpdatesResponse', describeServiceUpdatesResponse_serviceUpdates - A list of service updates

$sel:nextToken:DescribeServiceUpdates', describeServiceUpdatesResponse_nextToken - An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

$sel:httpStatus:DescribeServiceUpdatesResponse', describeServiceUpdatesResponse_httpStatus - The response's http status code.

Response Lenses

describeServiceUpdatesResponse_serviceUpdates :: Lens' DescribeServiceUpdatesResponse (Maybe [ServiceUpdate])

A list of service updates

describeServiceUpdatesResponse_nextToken :: Lens' DescribeServiceUpdatesResponse (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeServiceUpdatesResponse_httpStatus :: Lens' DescribeServiceUpdatesResponse Int

The response's http status code.