Amazonka.MemoryDb.DescribeSnapshots

Description

Returns information about cluster snapshots. By default, DescribeSnapshots lists all of your snapshots; it can optionally describe a single snapshot, or just the snapshots associated with a particular cluster.

Creating a Request

data DescribeSnapshots

See: newDescribeSnapshots smart constructor.

Constructors

DescribeSnapshots'

Fields

showDetail :: Maybe Bool

A Boolean value which if true, the shard configuration is included in the snapshot description.

nextToken :: Maybe Text

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

source :: Maybe Text

If set to system, the output shows snapshots that were automatically created by MemoryDB. If set to user the output shows snapshots that were manually created. If omitted, the output shows both automatically and manually created snapshots.

clusterName :: Maybe Text

A user-supplied cluster identifier. If this parameter is specified, only snapshots associated with that specific cluster are described.

snapshotName :: Maybe Text

A user-supplied name of the snapshot. If this parameter is specified, only this named snapshot is described.

maxResults :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

newDescribeSnapshots :: DescribeSnapshots

Create a value of DescribeSnapshots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:showDetail:DescribeSnapshots', describeSnapshots_showDetail - A Boolean value which if true, the shard configuration is included in the snapshot description.

$sel:nextToken:DescribeSnapshots', describeSnapshots_nextToken - An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

$sel:source:DescribeSnapshots', describeSnapshots_source - If set to system, the output shows snapshots that were automatically created by MemoryDB. If set to user the output shows snapshots that were manually created. If omitted, the output shows both automatically and manually created snapshots.

$sel:clusterName:DescribeSnapshots', describeSnapshots_clusterName - A user-supplied cluster identifier. If this parameter is specified, only snapshots associated with that specific cluster are described.

$sel:snapshotName:DescribeSnapshots', describeSnapshots_snapshotName - A user-supplied name of the snapshot. If this parameter is specified, only this named snapshot is described.

$sel:maxResults:DescribeSnapshots', describeSnapshots_maxResults - The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

Request Lenses

describeSnapshots_showDetail :: Lens' DescribeSnapshots (Maybe Bool)

A Boolean value which if true, the shard configuration is included in the snapshot description.

describeSnapshots_nextToken :: Lens' DescribeSnapshots (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeSnapshots_source :: Lens' DescribeSnapshots (Maybe Text)

If set to system, the output shows snapshots that were automatically created by MemoryDB. If set to user the output shows snapshots that were manually created. If omitted, the output shows both automatically and manually created snapshots.

describeSnapshots_clusterName :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied cluster identifier. If this parameter is specified, only snapshots associated with that specific cluster are described.

describeSnapshots_snapshotName :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied name of the snapshot. If this parameter is specified, only this named snapshot is described.

describeSnapshots_maxResults :: Lens' DescribeSnapshots (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

Destructuring the Response

data DescribeSnapshotsResponse

See: newDescribeSnapshotsResponse smart constructor.

Constructors

DescribeSnapshotsResponse'

Fields

nextToken :: Maybe Text

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

snapshots :: Maybe [Snapshot]

A list of snapshots. Each item in the list contains detailed information about one snapshot.

httpStatus :: Int

The response's http status code.

newDescribeSnapshotsResponse

Arguments

:: Int
$sel:httpStatus:DescribeSnapshotsResponse'

-> DescribeSnapshotsResponse

Create a value of DescribeSnapshotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSnapshots', describeSnapshotsResponse_nextToken - An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

$sel:snapshots:DescribeSnapshotsResponse', describeSnapshotsResponse_snapshots - A list of snapshots. Each item in the list contains detailed information about one snapshot.

$sel:httpStatus:DescribeSnapshotsResponse', describeSnapshotsResponse_httpStatus - The response's http status code.

Response Lenses

describeSnapshotsResponse_nextToken :: Lens' DescribeSnapshotsResponse (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeSnapshotsResponse_snapshots :: Lens' DescribeSnapshotsResponse (Maybe [Snapshot])

A list of snapshots. Each item in the list contains detailed information about one snapshot.

describeSnapshotsResponse_httpStatus :: Lens' DescribeSnapshotsResponse Int

The response's http status code.