Amazonka.MemoryDb.Lens

Description

Operations

DescribeClusters

describeClusters_showShardDetails :: Lens' DescribeClusters (Maybe Bool)

An optional flag that can be included in the request to retrieve information about the individual shard(s).

describeClusters_nextToken :: Lens' DescribeClusters (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeClusters_clusterName :: Lens' DescribeClusters (Maybe Text)

The name of the cluster

describeClusters_maxResults :: Lens' DescribeClusters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

describeClustersResponse_nextToken :: Lens' DescribeClustersResponse (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeClustersResponse_clusters :: Lens' DescribeClustersResponse (Maybe [Cluster])

A list of clusters

describeClustersResponse_httpStatus :: Lens' DescribeClustersResponse Int

The response's http status code.

BatchUpdateCluster

batchUpdateCluster_serviceUpdate :: Lens' BatchUpdateCluster (Maybe ServiceUpdateRequest)

The unique ID of the service update

batchUpdateCluster_clusterNames :: Lens' BatchUpdateCluster [Text]

The cluster names to apply the updates.

batchUpdateClusterResponse_unprocessedClusters :: Lens' BatchUpdateClusterResponse (Maybe [UnprocessedCluster])

The list of clusters where updates have not been applied.

batchUpdateClusterResponse_processedClusters :: Lens' BatchUpdateClusterResponse (Maybe [Cluster])

The list of clusters that have been updated.

batchUpdateClusterResponse_httpStatus :: Lens' BatchUpdateClusterResponse Int

The response's http status code.

DescribeUsers

describeUsers_filters :: Lens' DescribeUsers (Maybe [Filter])

Filter to determine the list of users to return.

describeUsers_userName :: Lens' DescribeUsers (Maybe Text)

The name of the user

describeUsers_nextToken :: Lens' DescribeUsers (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeUsers_maxResults :: Lens' DescribeUsers (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

describeUsersResponse_users :: Lens' DescribeUsersResponse (Maybe [User])

A list of users.

describeUsersResponse_nextToken :: Lens' DescribeUsersResponse (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeUsersResponse_httpStatus :: Lens' DescribeUsersResponse Int

The response's http status code.

DescribeParameters

describeParameters_nextToken :: Lens' DescribeParameters (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeParameters_maxResults :: Lens' DescribeParameters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

describeParameters_parameterGroupName :: Lens' DescribeParameters Text

he name of a specific parameter group to return details for.

describeParametersResponse_nextToken :: Lens' DescribeParametersResponse (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeParametersResponse_parameters :: Lens' DescribeParametersResponse (Maybe [Parameter])

A list of parameters specific to a particular parameter group. Each element in the list contains detailed information about one parameter.

describeParametersResponse_httpStatus :: Lens' DescribeParametersResponse Int

The response's http status code.

DeleteACL

deleteACL_aCLName :: Lens' DeleteACL Text

The name of the Access Control List to delete

deleteACLResponse_acl :: Lens' DeleteACLResponse (Maybe ACL)

The Access Control List object that has been deleted.

deleteACLResponse_httpStatus :: Lens' DeleteACLResponse Int

The response's http status code.

UpdateACL

updateACL_userNamesToAdd :: Lens' UpdateACL (Maybe (NonEmpty Text))

The list of users to add to the Access Control List

updateACL_userNamesToRemove :: Lens' UpdateACL (Maybe (NonEmpty Text))

The list of users to remove from the Access Control List

updateACL_aCLName :: Lens' UpdateACL Text

The name of the Access Control List

updateACLResponse_acl :: Lens' UpdateACLResponse (Maybe ACL)

The updated Access Control List

updateACLResponse_httpStatus :: Lens' UpdateACLResponse Int

The response's http status code.

DescribeEvents

describeEvents_sourceName :: Lens' DescribeEvents (Maybe Text)

The identifier of the event source for which events are returned. If not specified, all sources are included in the response.

describeEvents_startTime :: Lens' DescribeEvents (Maybe UTCTime)

The beginning of the time interval to retrieve events for, specified in ISO 8601 format. Example: 2017-03-30T07:03:49.555Z

describeEvents_sourceType :: Lens' DescribeEvents (Maybe SourceType)

The event source to retrieve events for. If no value is specified, all events are returned.

describeEvents_nextToken :: Lens' DescribeEvents (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeEvents_endTime :: Lens' DescribeEvents (Maybe UTCTime)

The end of the time interval for which to retrieve events, specified in ISO 8601 format. Example: 2017-03-30T07:03:49.555Z

describeEvents_duration :: Lens' DescribeEvents (Maybe Int)

The number of minutes worth of events to retrieve.

describeEvents_maxResults :: Lens' DescribeEvents (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

describeEventsResponse_nextToken :: Lens' DescribeEventsResponse (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeEventsResponse_events :: Lens' DescribeEventsResponse (Maybe [Event])

A list of events. Each element in the list contains detailed information about one event.

describeEventsResponse_httpStatus :: Lens' DescribeEventsResponse Int

The response's http status code.

DescribeEngineVersions

describeEngineVersions_engineVersion :: Lens' DescribeEngineVersions (Maybe Text)

The Redis engine version

describeEngineVersions_defaultOnly :: Lens' DescribeEngineVersions (Maybe Bool)

If true, specifies that only the default version of the specified engine or engine and major version combination is to be returned.

describeEngineVersions_nextToken :: Lens' DescribeEngineVersions (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeEngineVersions_parameterGroupFamily :: Lens' DescribeEngineVersions (Maybe Text)

The name of a specific parameter group family to return details for.

describeEngineVersions_maxResults :: Lens' DescribeEngineVersions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

describeEngineVersionsResponse_nextToken :: Lens' DescribeEngineVersionsResponse (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeEngineVersionsResponse_engineVersions :: Lens' DescribeEngineVersionsResponse (Maybe [EngineVersionInfo])

A list of engine version details. Each element in the list contains detailed information about one engine version.

describeEngineVersionsResponse_httpStatus :: Lens' DescribeEngineVersionsResponse Int

The response's http status code.

DescribeACLs

describeACLs_nextToken :: Lens' DescribeACLs (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeACLs_maxResults :: Lens' DescribeACLs (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

describeACLs_aCLName :: Lens' DescribeACLs (Maybe Text)

The name of the ACL

describeACLsResponse_nextToken :: Lens' DescribeACLsResponse (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeACLsResponse_aCLs :: Lens' DescribeACLsResponse (Maybe [ACL])

The list of ACLs

describeACLsResponse_httpStatus :: Lens' DescribeACLsResponse Int

The response's http status code.

CreateSubnetGroup

createSubnetGroup_description :: Lens' CreateSubnetGroup (Maybe Text)

A description for the subnet group.

createSubnetGroup_tags :: Lens' CreateSubnetGroup (Maybe [Tag])

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

createSubnetGroup_subnetGroupName :: Lens' CreateSubnetGroup Text

The name of the subnet group.

createSubnetGroup_subnetIds :: Lens' CreateSubnetGroup [Text]

A list of VPC subnet IDs for the subnet group.

createSubnetGroupResponse_subnetGroup :: Lens' CreateSubnetGroupResponse (Maybe SubnetGroup)

The newly-created subnet group

createSubnetGroupResponse_httpStatus :: Lens' CreateSubnetGroupResponse Int

The response's http status code.

DeleteCluster

deleteCluster_finalSnapshotName :: Lens' DeleteCluster (Maybe Text)

The user-supplied name of a final cluster snapshot. This is the unique name that identifies the snapshot. MemoryDB creates the snapshot, and then deletes the cluster immediately afterward.

deleteCluster_clusterName :: Lens' DeleteCluster Text

The name of the cluster to be deleted

deleteClusterResponse_cluster :: Lens' DeleteClusterResponse (Maybe Cluster)

The cluster object that has been deleted

deleteClusterResponse_httpStatus :: Lens' DeleteClusterResponse Int

The response's http status code.

UpdateCluster

updateCluster_engineVersion :: Lens' UpdateCluster (Maybe Text)

The upgraded version of the engine to be run on the nodes. You can upgrade to a newer engine version, but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cluster and create it anew with the earlier engine version.

updateCluster_securityGroupIds :: Lens' UpdateCluster (Maybe [Text])

The SecurityGroupIds to update

updateCluster_snsTopicArn :: Lens' UpdateCluster (Maybe Text)

The SNS topic ARN to update

updateCluster_snapshotWindow :: Lens' UpdateCluster (Maybe Text)

The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your cluster.

updateCluster_maintenanceWindow :: Lens' UpdateCluster (Maybe Text)

The maintenance window to update

updateCluster_shardConfiguration :: Lens' UpdateCluster (Maybe ShardConfigurationRequest)

The number of shards in the cluster

updateCluster_snapshotRetentionLimit :: Lens' UpdateCluster (Maybe Int)

The number of days for which MemoryDB retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

updateCluster_replicaConfiguration :: Lens' UpdateCluster (Maybe ReplicaConfigurationRequest)

The number of replicas that will reside in each shard

updateCluster_nodeType :: Lens' UpdateCluster (Maybe Text)

A valid node type that you want to scale this cluster up or down to.

updateCluster_snsTopicStatus :: Lens' UpdateCluster (Maybe Text)

The status of the Amazon SNS notification topic. Notifications are sent only if the status is active.

updateCluster_description :: Lens' UpdateCluster (Maybe Text)

The description of the cluster to update

updateCluster_aCLName :: Lens' UpdateCluster (Maybe Text)

The Access Control List that is associated with the cluster

updateCluster_parameterGroupName :: Lens' UpdateCluster (Maybe Text)

The name of the parameter group to update

updateCluster_clusterName :: Lens' UpdateCluster Text

The name of the cluster to update

updateClusterResponse_cluster :: Lens' UpdateClusterResponse (Maybe Cluster)

The updated cluster

updateClusterResponse_httpStatus :: Lens' UpdateClusterResponse Int

The response's http status code.

CreateCluster

createCluster_engineVersion :: Lens' CreateCluster (Maybe Text)

The version number of the Redis engine to be used for the cluster.

createCluster_securityGroupIds :: Lens' CreateCluster (Maybe [Text])

A list of security group names to associate with this cluster.

createCluster_snapshotArns :: Lens' CreateCluster (Maybe [Text])

A list of Amazon Resource Names (ARN) that uniquely identify the RDB snapshot files stored in Amazon S3. The snapshot files are used to populate the new cluster. The Amazon S3 object name in the ARN cannot contain any commas.

createCluster_autoMinorVersionUpgrade :: Lens' CreateCluster (Maybe Bool)

When set to true, the cluster will automatically receive minor engine version upgrades after launch.

createCluster_snsTopicArn :: Lens' CreateCluster (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications are sent.

createCluster_snapshotWindow :: Lens' CreateCluster (Maybe Text)

The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your shard.

Example: 05:00-09:00

If you do not specify this parameter, MemoryDB automatically chooses an appropriate time range.

createCluster_subnetGroupName :: Lens' CreateCluster (Maybe Text)

The name of the subnet group to be used for the cluster.

createCluster_tLSEnabled :: Lens' CreateCluster (Maybe Bool)

A flag to enable in-transit encryption on the cluster.

createCluster_numShards :: Lens' CreateCluster (Maybe Int)

The number of shards the cluster will contain. The default value is 1.

createCluster_maintenanceWindow :: Lens' CreateCluster (Maybe Text)

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

createCluster_kmsKeyId :: Lens' CreateCluster (Maybe Text)

The ID of the KMS key used to encrypt the cluster.

createCluster_snapshotRetentionLimit :: Lens' CreateCluster (Maybe Int)

The number of days for which MemoryDB retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

createCluster_snapshotName :: Lens' CreateCluster (Maybe Text)

The name of a snapshot from which to restore data into the new cluster. The snapshot status changes to restoring while the new cluster is being created.

createCluster_description :: Lens' CreateCluster (Maybe Text)

An optional description of the cluster.

createCluster_numReplicasPerShard :: Lens' CreateCluster (Maybe Int)

The number of replicas to apply to each shard. The default value is 1. The maximum is 5.

createCluster_tags :: Lens' CreateCluster (Maybe [Tag])

A list of tags to be added to this resource. Tags are comma-separated key,value pairs (e.g. Key=myKey, Value=myKeyValue. You can include multiple tags as shown following: Key=myKey, Value=myKeyValue Key=mySecondKey, Value=mySecondKeyValue.

createCluster_port :: Lens' CreateCluster (Maybe Int)

The port number on which each of the nodes accepts connections.

createCluster_parameterGroupName :: Lens' CreateCluster (Maybe Text)

The name of the parameter group associated with the cluster.

createCluster_clusterName :: Lens' CreateCluster Text

The name of the cluster. This value must be unique as it also serves as the cluster identifier.

createCluster_nodeType :: Lens' CreateCluster Text

The compute and memory capacity of the nodes in the cluster.

createCluster_aCLName :: Lens' CreateCluster Text

The name of the Access Control List to associate with the cluster.

createClusterResponse_cluster :: Lens' CreateClusterResponse (Maybe Cluster)

The newly-created cluster.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.

CopySnapshot

copySnapshot_targetBucket :: Lens' CopySnapshot (Maybe Text)

The Amazon S3 bucket to which the snapshot is exported. This parameter is used only when exporting a snapshot for external access. When using this parameter to export a snapshot, be sure MemoryDB has the needed permissions to this S3 bucket. For more information, see Step 2: Grant MemoryDB Access to Your Amazon S3 Bucket.

copySnapshot_kmsKeyId :: Lens' CopySnapshot (Maybe Text)

The ID of the KMS key used to encrypt the target snapshot.

copySnapshot_tags :: Lens' CopySnapshot (Maybe [Tag])

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

copySnapshot_sourceSnapshotName :: Lens' CopySnapshot Text

The name of an existing snapshot from which to make a copy.

copySnapshot_targetSnapshotName :: Lens' CopySnapshot Text

A name for the snapshot copy. MemoryDB does not permit overwriting a snapshot, therefore this name must be unique within its context - MemoryDB or an Amazon S3 bucket if exporting.

copySnapshotResponse_snapshot :: Lens' CopySnapshotResponse (Maybe Snapshot)

Represents a copy of an entire cluster as of the time when the snapshot was taken.

copySnapshotResponse_httpStatus :: Lens' CopySnapshotResponse Int

The response's http status code.

DeleteParameterGroup

deleteParameterGroup_parameterGroupName :: Lens' DeleteParameterGroup Text

The name of the parameter group to delete.

deleteParameterGroupResponse_parameterGroup :: Lens' DeleteParameterGroupResponse (Maybe ParameterGroup)

The parameter group that has been deleted.

deleteParameterGroupResponse_httpStatus :: Lens' DeleteParameterGroupResponse Int

The response's http status code.

UpdateParameterGroup

updateParameterGroup_parameterGroupName :: Lens' UpdateParameterGroup Text

The name of the parameter group to update.

updateParameterGroup_parameterNameValues :: Lens' UpdateParameterGroup [ParameterNameValue]

An array of parameter names and values for the parameter update. You must supply at least one parameter name and value; subsequent arguments are optional. A maximum of 20 parameters may be updated per request.

updateParameterGroupResponse_parameterGroup :: Lens' UpdateParameterGroupResponse (Maybe ParameterGroup)

The updated parameter group

updateParameterGroupResponse_httpStatus :: Lens' UpdateParameterGroupResponse Int

The response's http status code.

DescribeSubnetGroups

describeSubnetGroups_subnetGroupName :: Lens' DescribeSubnetGroups (Maybe Text)

The name of the subnet group to return details for.

describeSubnetGroups_nextToken :: Lens' DescribeSubnetGroups (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeSubnetGroups_maxResults :: Lens' DescribeSubnetGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

describeSubnetGroupsResponse_subnetGroups :: Lens' DescribeSubnetGroupsResponse (Maybe [SubnetGroup])

A list of subnet groups. Each element in the list contains detailed information about one group.

describeSubnetGroupsResponse_nextToken :: Lens' DescribeSubnetGroupsResponse (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeSubnetGroupsResponse_httpStatus :: Lens' DescribeSubnetGroupsResponse Int

The response's http status code.

DescribeServiceUpdates

describeServiceUpdates_status :: Lens' DescribeServiceUpdates (Maybe [ServiceUpdateStatus])

The status(es) of the service updates to filter on

describeServiceUpdates_serviceUpdateName :: Lens' DescribeServiceUpdates (Maybe Text)

The unique ID of the service update to describe.

describeServiceUpdates_clusterNames :: Lens' DescribeServiceUpdates (Maybe [Text])

The list of cluster names to identify service updates to apply

describeServiceUpdates_nextToken :: Lens' DescribeServiceUpdates (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeServiceUpdates_maxResults :: Lens' DescribeServiceUpdates (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

describeServiceUpdatesResponse_serviceUpdates :: Lens' DescribeServiceUpdatesResponse (Maybe [ServiceUpdate])

A list of service updates

describeServiceUpdatesResponse_nextToken :: Lens' DescribeServiceUpdatesResponse (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeServiceUpdatesResponse_httpStatus :: Lens' DescribeServiceUpdatesResponse Int

The response's http status code.

CreateParameterGroup

createParameterGroup_description :: Lens' CreateParameterGroup (Maybe Text)

An optional description of the parameter group.

createParameterGroup_tags :: Lens' CreateParameterGroup (Maybe [Tag])

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

createParameterGroup_parameterGroupName :: Lens' CreateParameterGroup Text

The name of the parameter group.

createParameterGroup_family :: Lens' CreateParameterGroup Text

The name of the parameter group family that the parameter group can be used with.

createParameterGroupResponse_parameterGroup :: Lens' CreateParameterGroupResponse (Maybe ParameterGroup)

The newly-created parameter group.

createParameterGroupResponse_httpStatus :: Lens' CreateParameterGroupResponse Int

The response's http status code.

DescribeSnapshots

describeSnapshots_showDetail :: Lens' DescribeSnapshots (Maybe Bool)

A Boolean value which if true, the shard configuration is included in the snapshot description.

describeSnapshots_nextToken :: Lens' DescribeSnapshots (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeSnapshots_source :: Lens' DescribeSnapshots (Maybe Text)

If set to system, the output shows snapshots that were automatically created by MemoryDB. If set to user the output shows snapshots that were manually created. If omitted, the output shows both automatically and manually created snapshots.

describeSnapshots_clusterName :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied cluster identifier. If this parameter is specified, only snapshots associated with that specific cluster are described.

describeSnapshots_snapshotName :: Lens' DescribeSnapshots (Maybe Text)

A user-supplied name of the snapshot. If this parameter is specified, only this named snapshot is described.

describeSnapshots_maxResults :: Lens' DescribeSnapshots (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

describeSnapshotsResponse_nextToken :: Lens' DescribeSnapshotsResponse (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeSnapshotsResponse_snapshots :: Lens' DescribeSnapshotsResponse (Maybe [Snapshot])

A list of snapshots. Each item in the list contains detailed information about one snapshot.

describeSnapshotsResponse_httpStatus :: Lens' DescribeSnapshotsResponse Int

The response's http status code.

CreateACL

createACL_userNames :: Lens' CreateACL (Maybe (NonEmpty Text))

The list of users that belong to the Access Control List.

createACL_tags :: Lens' CreateACL (Maybe [Tag])

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

createACL_aCLName :: Lens' CreateACL Text

The name of the Access Control List.

createACLResponse_acl :: Lens' CreateACLResponse (Maybe ACL)

The newly-created Access Control List.

createACLResponse_httpStatus :: Lens' CreateACLResponse Int

The response's http status code.

UpdateSubnetGroup

updateSubnetGroup_subnetIds :: Lens' UpdateSubnetGroup (Maybe [Text])

The EC2 subnet IDs for the subnet group.

updateSubnetGroup_description :: Lens' UpdateSubnetGroup (Maybe Text)

A description of the subnet group

updateSubnetGroup_subnetGroupName :: Lens' UpdateSubnetGroup Text

The name of the subnet group

updateSubnetGroupResponse_subnetGroup :: Lens' UpdateSubnetGroupResponse (Maybe SubnetGroup)

The updated subnet group

updateSubnetGroupResponse_httpStatus :: Lens' UpdateSubnetGroupResponse Int

The response's http status code.

DeleteSubnetGroup

deleteSubnetGroup_subnetGroupName :: Lens' DeleteSubnetGroup Text

The name of the subnet group to delete

deleteSubnetGroupResponse_subnetGroup :: Lens' DeleteSubnetGroupResponse (Maybe SubnetGroup)

The subnet group object that has been deleted.

deleteSubnetGroupResponse_httpStatus :: Lens' DeleteSubnetGroupResponse Int

The response's http status code.

CreateUser

createUser_tags :: Lens' CreateUser (Maybe [Tag])

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

createUser_userName :: Lens' CreateUser Text

The name of the user. This value must be unique as it also serves as the user identifier.

createUser_authenticationMode :: Lens' CreateUser AuthenticationMode

Denotes the user's authentication properties, such as whether it requires a password to authenticate.

createUser_accessString :: Lens' CreateUser Text

Access permissions string used for this user.

createUserResponse_user :: Lens' CreateUserResponse (Maybe User)

The newly-created user.

createUserResponse_httpStatus :: Lens' CreateUserResponse Int

The response's http status code.

DeleteSnapshot

deleteSnapshot_snapshotName :: Lens' DeleteSnapshot Text

The name of the snapshot to delete

deleteSnapshotResponse_snapshot :: Lens' DeleteSnapshotResponse (Maybe Snapshot)

The snapshot object that has been deleted.

deleteSnapshotResponse_httpStatus :: Lens' DeleteSnapshotResponse Int

The response's http status code.

FailoverShard

failoverShard_clusterName :: Lens' FailoverShard Text

The cluster being failed over

failoverShard_shardName :: Lens' FailoverShard Text

The name of the shard

failoverShardResponse_cluster :: Lens' FailoverShardResponse (Maybe Cluster)

The cluster being failed over

failoverShardResponse_httpStatus :: Lens' FailoverShardResponse Int

The response's http status code.

UpdateUser

updateUser_authenticationMode :: Lens' UpdateUser (Maybe AuthenticationMode)

Denotes the user's authentication properties, such as whether it requires a password to authenticate.

updateUser_accessString :: Lens' UpdateUser (Maybe Text)

Access permissions string used for this user.

updateUser_userName :: Lens' UpdateUser Text

The name of the user

updateUserResponse_user :: Lens' UpdateUserResponse (Maybe User)

The updated user

updateUserResponse_httpStatus :: Lens' UpdateUserResponse Int

The response's http status code.

DeleteUser

deleteUser_userName :: Lens' DeleteUser Text

The name of the user to delete

deleteUserResponse_user :: Lens' DeleteUserResponse (Maybe User)

The user object that has been deleted.

deleteUserResponse_httpStatus :: Lens' DeleteUserResponse Int

The response's http status code.

ListAllowedNodeTypeUpdates

listAllowedNodeTypeUpdates_clusterName :: Lens' ListAllowedNodeTypeUpdates Text

The name of the cluster you want to scale. MemoryDB uses the cluster name to identify the current node type being used by this cluster, and from that to create a list of node types you can scale up to.

listAllowedNodeTypeUpdatesResponse_scaleUpNodeTypes :: Lens' ListAllowedNodeTypeUpdatesResponse (Maybe [Text])

A list node types which you can use to scale up your cluster.

listAllowedNodeTypeUpdatesResponse_scaleDownNodeTypes :: Lens' ListAllowedNodeTypeUpdatesResponse (Maybe [Text])

A list node types which you can use to scale down your cluster.

listAllowedNodeTypeUpdatesResponse_httpStatus :: Lens' ListAllowedNodeTypeUpdatesResponse Int

The response's http status code.

DescribeParameterGroups

describeParameterGroups_nextToken :: Lens' DescribeParameterGroups (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeParameterGroups_maxResults :: Lens' DescribeParameterGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

describeParameterGroups_parameterGroupName :: Lens' DescribeParameterGroups (Maybe Text)

The name of a specific parameter group to return details for.

describeParameterGroupsResponse_nextToken :: Lens' DescribeParameterGroupsResponse (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeParameterGroupsResponse_parameterGroups :: Lens' DescribeParameterGroupsResponse (Maybe [ParameterGroup])

A list of parameter groups. Each element in the list contains detailed information about one parameter group.

describeParameterGroupsResponse_httpStatus :: Lens' DescribeParameterGroupsResponse Int

The response's http status code.

CreateSnapshot

createSnapshot_kmsKeyId :: Lens' CreateSnapshot (Maybe Text)

The ID of the KMS key used to encrypt the snapshot.

createSnapshot_tags :: Lens' CreateSnapshot (Maybe [Tag])

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

createSnapshot_clusterName :: Lens' CreateSnapshot Text

The snapshot is created from this cluster.

createSnapshot_snapshotName :: Lens' CreateSnapshot Text

A name for the snapshot being created.

createSnapshotResponse_snapshot :: Lens' CreateSnapshotResponse (Maybe Snapshot)

The newly-created snapshot.

createSnapshotResponse_httpStatus :: Lens' CreateSnapshotResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource to which the tags are to be added

tagResource_tags :: Lens' TagResource [Tag]

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

tagResourceResponse_tagList :: Lens' TagResourceResponse (Maybe [Tag])

A list of tags as key-value pairs.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListTags

listTags_resourceArn :: Lens' ListTags Text

The Amazon Resource Name (ARN) of the resource for which you want the list of tags

listTagsResponse_tagList :: Lens' ListTagsResponse (Maybe [Tag])

A list of tags as key-value pairs.

listTagsResponse_httpStatus :: Lens' ListTagsResponse Int

The response's http status code.

ResetParameterGroup

resetParameterGroup_parameterNames :: Lens' ResetParameterGroup (Maybe [Text])

An array of parameter names to reset to their default values. If AllParameters is true, do not use ParameterNames. If AllParameters is false, you must specify the name of at least one parameter to reset.

resetParameterGroup_allParameters :: Lens' ResetParameterGroup (Maybe Bool)

If true, all parameters in the parameter group are reset to their default values. If false, only the parameters listed by ParameterNames are reset to their default values.

resetParameterGroup_parameterGroupName :: Lens' ResetParameterGroup Text

The name of the parameter group to reset.

resetParameterGroupResponse_parameterGroup :: Lens' ResetParameterGroupResponse (Maybe ParameterGroup)

The parameter group being reset.

resetParameterGroupResponse_httpStatus :: Lens' ResetParameterGroupResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource to which the tags are to be removed

untagResource_tagKeys :: Lens' UntagResource [Text]

The list of keys of the tags that are to be removed

untagResourceResponse_tagList :: Lens' UntagResourceResponse (Maybe [Tag])

The list of tags removed

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

Types

ACL

acl_status :: Lens' ACL (Maybe Text)

Indicates ACL status. Can be "creating", "active", "modifying", "deleting".

acl_userNames :: Lens' ACL (Maybe [Text])

The list of user names that belong to the ACL.

acl_arn :: Lens' ACL (Maybe Text)

The Amazon Resource Name (ARN) of the ACL

acl_name :: Lens' ACL (Maybe Text)

The name of the Access Control List

acl_pendingChanges :: Lens' ACL (Maybe ACLPendingChanges)

A list of updates being applied to the ACL.

acl_minimumEngineVersion :: Lens' ACL (Maybe Text)

The minimum engine version supported for the ACL

acl_clusters :: Lens' ACL (Maybe [Text])

A list of clusters associated with the ACL.

ACLPendingChanges

aCLPendingChanges_userNamesToAdd :: Lens' ACLPendingChanges (Maybe [Text])

A list of users being added to the ACL

aCLPendingChanges_userNamesToRemove :: Lens' ACLPendingChanges (Maybe [Text])

A list of user names being removed from the ACL

ACLsUpdateStatus

aCLsUpdateStatus_aCLToApply :: Lens' ACLsUpdateStatus (Maybe Text)

A list of ACLs pending to be applied.

Authentication

authentication_passwordCount :: Lens' Authentication (Maybe Int)

The number of passwords belonging to the user. The maximum is two.

authentication_type :: Lens' Authentication (Maybe AuthenticationType)

Indicates whether the user requires a password to authenticate.

AuthenticationMode

authenticationMode_passwords :: Lens' AuthenticationMode (Maybe (NonEmpty Text))

The password(s) used for authentication

authenticationMode_type :: Lens' AuthenticationMode (Maybe InputAuthenticationType)

Indicates whether the user requires a password to authenticate. All newly-created users require a password.

AvailabilityZone

availabilityZone_name :: Lens' AvailabilityZone (Maybe Text)

The name of the Availability Zone.

Cluster

cluster_engineVersion :: Lens' Cluster (Maybe Text)

The Redis engine version used by the cluster

cluster_status :: Lens' Cluster (Maybe Text)

The status of the cluster. For example, Available, Updating, Creating.

cluster_autoMinorVersionUpgrade :: Lens' Cluster (Maybe Bool)

When set to true, the cluster will automatically receive minor engine version upgrades after launch.

cluster_snsTopicArn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) of the SNS notification topic

cluster_securityGroups :: Lens' Cluster (Maybe [SecurityGroupMembership])

A list of security groups used by the cluster

cluster_availabilityMode :: Lens' Cluster (Maybe AZStatus)

Indicates if the cluster has a Multi-AZ configuration (multiaz) or not (singleaz).

cluster_parameterGroupStatus :: Lens' Cluster (Maybe Text)

The status of the parameter group used by the cluster, for example 'active' or 'applying'.

cluster_arn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

cluster_pendingUpdates :: Lens' Cluster (Maybe ClusterPendingUpdates)

A group of settings that are currently being applied.

cluster_numberOfShards :: Lens' Cluster (Maybe Int)

The number of shards in the cluster

cluster_snapshotWindow :: Lens' Cluster (Maybe Text)

The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your shard. Example: 05:00-09:00 If you do not specify this parameter, MemoryDB automatically chooses an appropriate time range.

cluster_subnetGroupName :: Lens' Cluster (Maybe Text)

The name of the subnet group used by the cluster

cluster_tLSEnabled :: Lens' Cluster (Maybe Bool)

A flag to indicate if In-transit encryption is enabled

cluster_maintenanceWindow :: Lens' Cluster (Maybe Text)

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

cluster_kmsKeyId :: Lens' Cluster (Maybe Text)

The ID of the KMS key used to encrypt the cluster

cluster_shards :: Lens' Cluster (Maybe [Shard])

A list of shards that are members of the cluster.

cluster_name :: Lens' Cluster (Maybe Text)

The user-supplied name of the cluster. This identifier is a unique key that identifies a cluster.

cluster_enginePatchVersion :: Lens' Cluster (Maybe Text)

The Redis engine patch version used by the cluster

cluster_snapshotRetentionLimit :: Lens' Cluster (Maybe Int)

The number of days for which MemoryDB retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

cluster_nodeType :: Lens' Cluster (Maybe Text)

The cluster's node type

cluster_snsTopicStatus :: Lens' Cluster (Maybe Text)

The SNS topic must be in Active status to receive notifications

cluster_description :: Lens' Cluster (Maybe Text)

A description of the cluster

cluster_aCLName :: Lens' Cluster (Maybe Text)

The name of the Access Control List associated with this cluster.

cluster_clusterEndpoint :: Lens' Cluster (Maybe Endpoint)

The cluster's configuration endpoint

cluster_parameterGroupName :: Lens' Cluster (Maybe Text)

The name of the parameter group used by the cluster

ClusterConfiguration

clusterConfiguration_engineVersion :: Lens' ClusterConfiguration (Maybe Text)

The Redis engine version used by the cluster

clusterConfiguration_vpcId :: Lens' ClusterConfiguration (Maybe Text)

The ID of the VPC the cluster belongs to

clusterConfiguration_snapshotWindow :: Lens' ClusterConfiguration (Maybe Text)

The snapshot window set by the cluster

clusterConfiguration_subnetGroupName :: Lens' ClusterConfiguration (Maybe Text)

The name of the subnet group used by the cluster

clusterConfiguration_numShards :: Lens' ClusterConfiguration (Maybe Int)

The number of shards in the cluster

clusterConfiguration_maintenanceWindow :: Lens' ClusterConfiguration (Maybe Text)

The specified maintenance window for the cluster

clusterConfiguration_topicArn :: Lens' ClusterConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the SNS notification topic for the cluster

clusterConfiguration_shards :: Lens' ClusterConfiguration (Maybe [ShardDetail])

The list of shards in the cluster

clusterConfiguration_name :: Lens' ClusterConfiguration (Maybe Text)

The name of the cluster

clusterConfiguration_snapshotRetentionLimit :: Lens' ClusterConfiguration (Maybe Int)

The snapshot retention limit set by the cluster

clusterConfiguration_nodeType :: Lens' ClusterConfiguration (Maybe Text)

The node type used for the cluster

clusterConfiguration_description :: Lens' ClusterConfiguration (Maybe Text)

The description of the cluster configuration

clusterConfiguration_port :: Lens' ClusterConfiguration (Maybe Int)

The port used by the cluster

clusterConfiguration_parameterGroupName :: Lens' ClusterConfiguration (Maybe Text)

The name of parameter group used by the cluster

ClusterPendingUpdates

clusterPendingUpdates_serviceUpdates :: Lens' ClusterPendingUpdates (Maybe [PendingModifiedServiceUpdate])

A list of service updates being applied to the cluster

clusterPendingUpdates_resharding :: Lens' ClusterPendingUpdates (Maybe ReshardingStatus)

The status of an online resharding operation.

clusterPendingUpdates_aCLs :: Lens' ClusterPendingUpdates (Maybe ACLsUpdateStatus)

A list of ACLs associated with the cluster that are being updated

Endpoint

endpoint_address :: Lens' Endpoint (Maybe Text)

The DNS hostname of the node.

endpoint_port :: Lens' Endpoint (Maybe Int)

The port number that the engine is listening on.

EngineVersionInfo

engineVersionInfo_engineVersion :: Lens' EngineVersionInfo (Maybe Text)

The engine version

engineVersionInfo_enginePatchVersion :: Lens' EngineVersionInfo (Maybe Text)

The patched engine version

engineVersionInfo_parameterGroupFamily :: Lens' EngineVersionInfo (Maybe Text)

Specifies the name of the parameter group family to which the engine default parameters apply.

Event

event_sourceName :: Lens' Event (Maybe Text)

The name for the source of the event. For example, if the event occurred at the cluster level, the identifier would be the name of the cluster.

event_sourceType :: Lens' Event (Maybe SourceType)

Specifies the origin of this event - a cluster, a parameter group, a security group, etc.

event_date :: Lens' Event (Maybe UTCTime)

The date and time when the event occurred.

event_message :: Lens' Event (Maybe Text)

The text of the event.

Filter

filter_name :: Lens' Filter Text

The property being filtered. For example, UserName.

filter_values :: Lens' Filter (NonEmpty Text)

The property values to filter on. For example, "user-123".

Node

node_status :: Lens' Node (Maybe Text)

The status of the service update on the node

node_availabilityZone :: Lens' Node (Maybe Text)

The Availability Zone in which the node resides

node_name :: Lens' Node (Maybe Text)

The node identifier. A node name is a numeric identifier (0001, 0002, etc.). The combination of cluster name, shard name and node name uniquely identifies every node used in a customer's Amazon account.

node_endpoint :: Lens' Node (Maybe Endpoint)

The hostname for connecting to this node.

node_createTime :: Lens' Node (Maybe UTCTime)

The date and time when the node was created.

Parameter

parameter_value :: Lens' Parameter (Maybe Text)

The value of the parameter

parameter_name :: Lens' Parameter (Maybe Text)

The name of the parameter

parameter_minimumEngineVersion :: Lens' Parameter (Maybe Text)

The earliest engine version to which the parameter can apply.

parameter_dataType :: Lens' Parameter (Maybe Text)

The parameter's data type

parameter_allowedValues :: Lens' Parameter (Maybe Text)

The valid range of values for the parameter.

parameter_description :: Lens' Parameter (Maybe Text)

A description of the parameter

ParameterGroup

parameterGroup_arn :: Lens' ParameterGroup (Maybe Text)

The Amazon Resource Name (ARN) of the parameter group

parameterGroup_family :: Lens' ParameterGroup (Maybe Text)

The name of the parameter group family that this parameter group is compatible with.

parameterGroup_name :: Lens' ParameterGroup (Maybe Text)

The name of the parameter group

parameterGroup_description :: Lens' ParameterGroup (Maybe Text)

A description of the parameter group

ParameterNameValue

parameterNameValue_parameterValue :: Lens' ParameterNameValue (Maybe Text)

The value of the parameter

parameterNameValue_parameterName :: Lens' ParameterNameValue (Maybe Text)

The name of the parameter

PendingModifiedServiceUpdate

pendingModifiedServiceUpdate_status :: Lens' PendingModifiedServiceUpdate (Maybe ServiceUpdateStatus)

The status of the service update

pendingModifiedServiceUpdate_serviceUpdateName :: Lens' PendingModifiedServiceUpdate (Maybe Text)

The unique ID of the service update

ReplicaConfigurationRequest

replicaConfigurationRequest_replicaCount :: Lens' ReplicaConfigurationRequest (Maybe Int)

The number of replicas to scale up or down to

ReshardingStatus

reshardingStatus_slotMigration :: Lens' ReshardingStatus (Maybe SlotMigration)

The status of the online resharding slot migration

SecurityGroupMembership

securityGroupMembership_status :: Lens' SecurityGroupMembership (Maybe Text)

The status of the security group membership. The status changes whenever a security group is modified, or when the security groups assigned to a cluster are modified.

securityGroupMembership_securityGroupId :: Lens' SecurityGroupMembership (Maybe Text)

The identifier of the security group.

ServiceUpdate

serviceUpdate_status :: Lens' ServiceUpdate (Maybe ServiceUpdateStatus)

The status of the service update

serviceUpdate_autoUpdateStartDate :: Lens' ServiceUpdate (Maybe UTCTime)

The date at which the service update will be automatically applied

serviceUpdate_serviceUpdateName :: Lens' ServiceUpdate (Maybe Text)

The unique ID of the service update

serviceUpdate_nodesUpdated :: Lens' ServiceUpdate (Maybe Text)

A list of nodes updated by the service update

serviceUpdate_releaseDate :: Lens' ServiceUpdate (Maybe UTCTime)

The date when the service update is initially available

serviceUpdate_clusterName :: Lens' ServiceUpdate (Maybe Text)

The name of the cluster to which the service update applies

serviceUpdate_type :: Lens' ServiceUpdate (Maybe ServiceUpdateType)

Reflects the nature of the service update

serviceUpdate_description :: Lens' ServiceUpdate (Maybe Text)

Provides details of the service update

ServiceUpdateRequest

serviceUpdateRequest_serviceUpdateNameToApply :: Lens' ServiceUpdateRequest (Maybe Text)

The unique ID of the service update

Shard

shard_status :: Lens' Shard (Maybe Text)

The current state of this replication group - creating, available, modifying, deleting.

shard_slots :: Lens' Shard (Maybe Text)

The keyspace for this shard.

shard_numberOfNodes :: Lens' Shard (Maybe Int)

The number of nodes in the shard

shard_name :: Lens' Shard (Maybe Text)

The name of the shard

shard_nodes :: Lens' Shard (Maybe [Node])

A list containing information about individual nodes within the shard

ShardConfiguration

shardConfiguration_slots :: Lens' ShardConfiguration (Maybe Text)

A string that specifies the keyspace for a particular node group. Keyspaces range from 0 to 16,383. The string is in the format startkey-endkey.

shardConfiguration_replicaCount :: Lens' ShardConfiguration (Maybe Int)

The number of read replica nodes in this shard.

ShardConfigurationRequest

shardConfigurationRequest_shardCount :: Lens' ShardConfigurationRequest (Maybe Int)

The number of shards in the cluster

ShardDetail

shardDetail_size :: Lens' ShardDetail (Maybe Text)

The size of the shard's snapshot

shardDetail_snapshotCreationTime :: Lens' ShardDetail (Maybe UTCTime)

The date and time that the shard's snapshot was created

shardDetail_name :: Lens' ShardDetail (Maybe Text)

The name of the shard

shardDetail_configuration :: Lens' ShardDetail (Maybe ShardConfiguration)

The configuration details of the shard

SlotMigration

slotMigration_progressPercentage :: Lens' SlotMigration (Maybe Double)

The percentage of the slot migration that is complete.

Snapshot

snapshot_status :: Lens' Snapshot (Maybe Text)

The status of the snapshot. Valid values: creating | available | restoring | copying | deleting.

snapshot_arn :: Lens' Snapshot (Maybe Text)

The ARN (Amazon Resource Name) of the snapshot.

snapshot_kmsKeyId :: Lens' Snapshot (Maybe Text)

The ID of the KMS key used to encrypt the snapshot.

snapshot_name :: Lens' Snapshot (Maybe Text)

The name of the snapshot

snapshot_clusterConfiguration :: Lens' Snapshot (Maybe ClusterConfiguration)

The configuration of the cluster from which the snapshot was taken

snapshot_source :: Lens' Snapshot (Maybe Text)

Indicates whether the snapshot is from an automatic backup (automated) or was created manually (manual).

Subnet

subnet_identifier :: Lens' Subnet (Maybe Text)

The unique identifier for the subnet.

subnet_availabilityZone :: Lens' Subnet (Maybe AvailabilityZone)

The Availability Zone where the subnet resides

SubnetGroup

subnetGroup_arn :: Lens' SubnetGroup (Maybe Text)

The ARN (Amazon Resource Name) of the subnet group.

subnetGroup_vpcId :: Lens' SubnetGroup (Maybe Text)

The Amazon Virtual Private Cloud identifier (VPC ID) of the subnet group.

subnetGroup_subnets :: Lens' SubnetGroup (Maybe [Subnet])

A list of subnets associated with the subnet group.

subnetGroup_name :: Lens' SubnetGroup (Maybe Text)

The name of the subnet group

subnetGroup_description :: Lens' SubnetGroup (Maybe Text)

A description of the subnet group

Tag

tag_value :: Lens' Tag (Maybe Text)

The tag's value. May be null.

tag_key :: Lens' Tag (Maybe Text)

The key for the tag. May not be null.

UnprocessedCluster

unprocessedCluster_clusterName :: Lens' UnprocessedCluster (Maybe Text)

The name of the cluster

unprocessedCluster_errorType :: Lens' UnprocessedCluster (Maybe Text)

The error type associated with the update failure

unprocessedCluster_errorMessage :: Lens' UnprocessedCluster (Maybe Text)

The error message associated with the update failure

User

user_status :: Lens' User (Maybe Text)

Indicates the user status. Can be "active", "modifying" or "deleting".

user_arn :: Lens' User (Maybe Text)

The Amazon Resource Name (ARN) of the user.

user_authentication :: Lens' User (Maybe Authentication)

Denotes whether the user requires a password to authenticate.

user_accessString :: Lens' User (Maybe Text)

Access permissions string used for this user.

user_name :: Lens' User (Maybe Text)

The name of the user

user_aCLNames :: Lens' User (Maybe [Text])

The names of the Access Control Lists to which the user belongs

user_minimumEngineVersion :: Lens' User (Maybe Text)

The minimum engine version supported for the user