Amazonka.MemoryDb.Types.Cluster

Description

Documentation

data Cluster

Contains all of the attributes of a specific cluster.

See: newCluster smart constructor.

Constructors

Cluster'

Fields

engineVersion :: Maybe Text

The Redis engine version used by the cluster

status :: Maybe Text

The status of the cluster. For example, Available, Updating, Creating.

autoMinorVersionUpgrade :: Maybe Bool

When set to true, the cluster will automatically receive minor engine version upgrades after launch.

snsTopicArn :: Maybe Text

The Amazon Resource Name (ARN) of the SNS notification topic

securityGroups :: Maybe [SecurityGroupMembership]

A list of security groups used by the cluster

availabilityMode :: Maybe AZStatus

Indicates if the cluster has a Multi-AZ configuration (multiaz) or not (singleaz).

parameterGroupStatus :: Maybe Text

The status of the parameter group used by the cluster, for example 'active' or 'applying'.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster.

pendingUpdates :: Maybe ClusterPendingUpdates

A group of settings that are currently being applied.

numberOfShards :: Maybe Int

The number of shards in the cluster

snapshotWindow :: Maybe Text

The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your shard. Example: 05:00-09:00 If you do not specify this parameter, MemoryDB automatically chooses an appropriate time range.

subnetGroupName :: Maybe Text

The name of the subnet group used by the cluster

tLSEnabled :: Maybe Bool

A flag to indicate if In-transit encryption is enabled

maintenanceWindow :: Maybe Text

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

kmsKeyId :: Maybe Text

The ID of the KMS key used to encrypt the cluster

shards :: Maybe [Shard]

A list of shards that are members of the cluster.

name :: Maybe Text

The user-supplied name of the cluster. This identifier is a unique key that identifies a cluster.

enginePatchVersion :: Maybe Text

The Redis engine patch version used by the cluster

snapshotRetentionLimit :: Maybe Int

The number of days for which MemoryDB retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

nodeType :: Maybe Text

The cluster's node type

snsTopicStatus :: Maybe Text

The SNS topic must be in Active status to receive notifications

description :: Maybe Text

A description of the cluster

aCLName :: Maybe Text

The name of the Access Control List associated with this cluster.

clusterEndpoint :: Maybe Endpoint

The cluster's configuration endpoint

parameterGroupName :: Maybe Text

The name of the parameter group used by the cluster

newCluster :: Cluster

Create a value of Cluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:Cluster', cluster_engineVersion - The Redis engine version used by the cluster

$sel:status:Cluster', cluster_status - The status of the cluster. For example, Available, Updating, Creating.

$sel:autoMinorVersionUpgrade:Cluster', cluster_autoMinorVersionUpgrade - When set to true, the cluster will automatically receive minor engine version upgrades after launch.

$sel:snsTopicArn:Cluster', cluster_snsTopicArn - The Amazon Resource Name (ARN) of the SNS notification topic

$sel:securityGroups:Cluster', cluster_securityGroups - A list of security groups used by the cluster

$sel:availabilityMode:Cluster', cluster_availabilityMode - Indicates if the cluster has a Multi-AZ configuration (multiaz) or not (singleaz).

$sel:parameterGroupStatus:Cluster', cluster_parameterGroupStatus - The status of the parameter group used by the cluster, for example 'active' or 'applying'.

$sel:arn:Cluster', cluster_arn - The Amazon Resource Name (ARN) of the cluster.

$sel:pendingUpdates:Cluster', cluster_pendingUpdates - A group of settings that are currently being applied.

$sel:numberOfShards:Cluster', cluster_numberOfShards - The number of shards in the cluster

$sel:snapshotWindow:Cluster', cluster_snapshotWindow - The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your shard. Example: 05:00-09:00 If you do not specify this parameter, MemoryDB automatically chooses an appropriate time range.

$sel:subnetGroupName:Cluster', cluster_subnetGroupName - The name of the subnet group used by the cluster

$sel:tLSEnabled:Cluster', cluster_tLSEnabled - A flag to indicate if In-transit encryption is enabled

$sel:maintenanceWindow:Cluster', cluster_maintenanceWindow - Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

$sel:kmsKeyId:Cluster', cluster_kmsKeyId - The ID of the KMS key used to encrypt the cluster

$sel:shards:Cluster', cluster_shards - A list of shards that are members of the cluster.

$sel:name:Cluster', cluster_name - The user-supplied name of the cluster. This identifier is a unique key that identifies a cluster.

$sel:enginePatchVersion:Cluster', cluster_enginePatchVersion - The Redis engine patch version used by the cluster

$sel:snapshotRetentionLimit:Cluster', cluster_snapshotRetentionLimit - The number of days for which MemoryDB retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

$sel:nodeType:Cluster', cluster_nodeType - The cluster's node type

$sel:snsTopicStatus:Cluster', cluster_snsTopicStatus - The SNS topic must be in Active status to receive notifications

$sel:description:Cluster', cluster_description - A description of the cluster

$sel:aCLName:Cluster', cluster_aCLName - The name of the Access Control List associated with this cluster.

$sel:clusterEndpoint:Cluster', cluster_clusterEndpoint - The cluster's configuration endpoint

$sel:parameterGroupName:Cluster', cluster_parameterGroupName - The name of the parameter group used by the cluster

cluster_engineVersion :: Lens' Cluster (Maybe Text)

The Redis engine version used by the cluster

cluster_status :: Lens' Cluster (Maybe Text)

The status of the cluster. For example, Available, Updating, Creating.

cluster_autoMinorVersionUpgrade :: Lens' Cluster (Maybe Bool)

When set to true, the cluster will automatically receive minor engine version upgrades after launch.

cluster_snsTopicArn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) of the SNS notification topic

cluster_securityGroups :: Lens' Cluster (Maybe [SecurityGroupMembership])

A list of security groups used by the cluster

cluster_availabilityMode :: Lens' Cluster (Maybe AZStatus)

Indicates if the cluster has a Multi-AZ configuration (multiaz) or not (singleaz).

cluster_parameterGroupStatus :: Lens' Cluster (Maybe Text)

The status of the parameter group used by the cluster, for example 'active' or 'applying'.

cluster_arn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

cluster_pendingUpdates :: Lens' Cluster (Maybe ClusterPendingUpdates)

A group of settings that are currently being applied.

cluster_numberOfShards :: Lens' Cluster (Maybe Int)

The number of shards in the cluster

cluster_snapshotWindow :: Lens' Cluster (Maybe Text)

The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your shard. Example: 05:00-09:00 If you do not specify this parameter, MemoryDB automatically chooses an appropriate time range.

cluster_subnetGroupName :: Lens' Cluster (Maybe Text)

The name of the subnet group used by the cluster

cluster_tLSEnabled :: Lens' Cluster (Maybe Bool)

A flag to indicate if In-transit encryption is enabled

cluster_maintenanceWindow :: Lens' Cluster (Maybe Text)

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

cluster_kmsKeyId :: Lens' Cluster (Maybe Text)

The ID of the KMS key used to encrypt the cluster

cluster_shards :: Lens' Cluster (Maybe [Shard])

A list of shards that are members of the cluster.

cluster_name :: Lens' Cluster (Maybe Text)

The user-supplied name of the cluster. This identifier is a unique key that identifies a cluster.

cluster_enginePatchVersion :: Lens' Cluster (Maybe Text)

The Redis engine patch version used by the cluster

cluster_snapshotRetentionLimit :: Lens' Cluster (Maybe Int)

The number of days for which MemoryDB retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

cluster_nodeType :: Lens' Cluster (Maybe Text)

The cluster's node type

cluster_snsTopicStatus :: Lens' Cluster (Maybe Text)

The SNS topic must be in Active status to receive notifications

cluster_description :: Lens' Cluster (Maybe Text)

A description of the cluster

cluster_aCLName :: Lens' Cluster (Maybe Text)

The name of the Access Control List associated with this cluster.

cluster_clusterEndpoint :: Lens' Cluster (Maybe Endpoint)

The cluster's configuration endpoint

cluster_parameterGroupName :: Lens' Cluster (Maybe Text)

The name of the parameter group used by the cluster