Amazonka.MemoryDb.Types.ServiceUpdate

Description

Documentation

data ServiceUpdate

An update that you can apply to your MemoryDB clusters.

See: newServiceUpdate smart constructor.

Constructors

ServiceUpdate'

Fields

status :: Maybe ServiceUpdateStatus

The status of the service update

autoUpdateStartDate :: Maybe POSIX

The date at which the service update will be automatically applied

serviceUpdateName :: Maybe Text

The unique ID of the service update

nodesUpdated :: Maybe Text

A list of nodes updated by the service update

releaseDate :: Maybe POSIX

The date when the service update is initially available

clusterName :: Maybe Text

The name of the cluster to which the service update applies

type' :: Maybe ServiceUpdateType

Reflects the nature of the service update

description :: Maybe Text

Provides details of the service update

newServiceUpdate :: ServiceUpdate

Create a value of ServiceUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ServiceUpdate', serviceUpdate_status - The status of the service update

$sel:autoUpdateStartDate:ServiceUpdate', serviceUpdate_autoUpdateStartDate - The date at which the service update will be automatically applied

$sel:serviceUpdateName:ServiceUpdate', serviceUpdate_serviceUpdateName - The unique ID of the service update

$sel:nodesUpdated:ServiceUpdate', serviceUpdate_nodesUpdated - A list of nodes updated by the service update

$sel:releaseDate:ServiceUpdate', serviceUpdate_releaseDate - The date when the service update is initially available

$sel:clusterName:ServiceUpdate', serviceUpdate_clusterName - The name of the cluster to which the service update applies

$sel:type':ServiceUpdate', serviceUpdate_type - Reflects the nature of the service update

$sel:description:ServiceUpdate', serviceUpdate_description - Provides details of the service update

serviceUpdate_status :: Lens' ServiceUpdate (Maybe ServiceUpdateStatus)

The status of the service update

serviceUpdate_autoUpdateStartDate :: Lens' ServiceUpdate (Maybe UTCTime)

The date at which the service update will be automatically applied

serviceUpdate_serviceUpdateName :: Lens' ServiceUpdate (Maybe Text)

The unique ID of the service update

serviceUpdate_nodesUpdated :: Lens' ServiceUpdate (Maybe Text)

A list of nodes updated by the service update

serviceUpdate_releaseDate :: Lens' ServiceUpdate (Maybe UTCTime)

The date when the service update is initially available

serviceUpdate_clusterName :: Lens' ServiceUpdate (Maybe Text)

The name of the cluster to which the service update applies

serviceUpdate_type :: Lens' ServiceUpdate (Maybe ServiceUpdateType)

Reflects the nature of the service update

serviceUpdate_description :: Lens' ServiceUpdate (Maybe Text)

Provides details of the service update