Amazonka.MemoryDb.Types.Snapshot

Description

Documentation

data Snapshot

Represents a copy of an entire cluster as of the time when the snapshot was taken.

See: newSnapshot smart constructor.

Constructors

Snapshot'

Fields

status :: Maybe Text

The status of the snapshot. Valid values: creating | available | restoring | copying | deleting.

arn :: Maybe Text

The ARN (Amazon Resource Name) of the snapshot.

kmsKeyId :: Maybe Text

The ID of the KMS key used to encrypt the snapshot.

name :: Maybe Text

The name of the snapshot

clusterConfiguration :: Maybe ClusterConfiguration

The configuration of the cluster from which the snapshot was taken

source :: Maybe Text

Indicates whether the snapshot is from an automatic backup (automated) or was created manually (manual).

newSnapshot :: Snapshot

Create a value of Snapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Snapshot', snapshot_status - The status of the snapshot. Valid values: creating | available | restoring | copying | deleting.

$sel:arn:Snapshot', snapshot_arn - The ARN (Amazon Resource Name) of the snapshot.

$sel:kmsKeyId:Snapshot', snapshot_kmsKeyId - The ID of the KMS key used to encrypt the snapshot.

$sel:name:Snapshot', snapshot_name - The name of the snapshot

$sel:clusterConfiguration:Snapshot', snapshot_clusterConfiguration - The configuration of the cluster from which the snapshot was taken

$sel:source:Snapshot', snapshot_source - Indicates whether the snapshot is from an automatic backup (automated) or was created manually (manual).

snapshot_status :: Lens' Snapshot (Maybe Text)

The status of the snapshot. Valid values: creating | available | restoring | copying | deleting.

snapshot_arn :: Lens' Snapshot (Maybe Text)

The ARN (Amazon Resource Name) of the snapshot.

snapshot_kmsKeyId :: Lens' Snapshot (Maybe Text)

The ID of the KMS key used to encrypt the snapshot.

snapshot_name :: Lens' Snapshot (Maybe Text)

The name of the snapshot

snapshot_clusterConfiguration :: Lens' Snapshot (Maybe ClusterConfiguration)

The configuration of the cluster from which the snapshot was taken

snapshot_source :: Lens' Snapshot (Maybe Text)

Indicates whether the snapshot is from an automatic backup (automated) or was created manually (manual).