Amazonka.MemoryDb.UpdateCluster

Description

Modifies the settings for a cluster. You can use this operation to change one or more cluster configuration settings by specifying the settings and the new values.

Creating a Request

data UpdateCluster

See: newUpdateCluster smart constructor.

Constructors

UpdateCluster'

Fields

engineVersion :: Maybe Text

The upgraded version of the engine to be run on the nodes. You can upgrade to a newer engine version, but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cluster and create it anew with the earlier engine version.

securityGroupIds :: Maybe [Text]

The SecurityGroupIds to update

snsTopicArn :: Maybe Text

The SNS topic ARN to update

snapshotWindow :: Maybe Text

The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your cluster.

maintenanceWindow :: Maybe Text

The maintenance window to update

shardConfiguration :: Maybe ShardConfigurationRequest

The number of shards in the cluster

snapshotRetentionLimit :: Maybe Int

The number of days for which MemoryDB retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

replicaConfiguration :: Maybe ReplicaConfigurationRequest

The number of replicas that will reside in each shard

nodeType :: Maybe Text

A valid node type that you want to scale this cluster up or down to.

snsTopicStatus :: Maybe Text

The status of the Amazon SNS notification topic. Notifications are sent only if the status is active.

description :: Maybe Text

The description of the cluster to update

aCLName :: Maybe Text

The Access Control List that is associated with the cluster

parameterGroupName :: Maybe Text

The name of the parameter group to update

clusterName :: Text

The name of the cluster to update

newUpdateCluster

Arguments

:: Text
$sel:clusterName:UpdateCluster'

-> UpdateCluster

Create a value of UpdateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:UpdateCluster', updateCluster_engineVersion - The upgraded version of the engine to be run on the nodes. You can upgrade to a newer engine version, but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cluster and create it anew with the earlier engine version.

$sel:securityGroupIds:UpdateCluster', updateCluster_securityGroupIds - The SecurityGroupIds to update

$sel:snsTopicArn:UpdateCluster', updateCluster_snsTopicArn - The SNS topic ARN to update

$sel:snapshotWindow:UpdateCluster', updateCluster_snapshotWindow - The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your cluster.

$sel:maintenanceWindow:UpdateCluster', updateCluster_maintenanceWindow - The maintenance window to update

$sel:shardConfiguration:UpdateCluster', updateCluster_shardConfiguration - The number of shards in the cluster

$sel:snapshotRetentionLimit:UpdateCluster', updateCluster_snapshotRetentionLimit - The number of days for which MemoryDB retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

$sel:replicaConfiguration:UpdateCluster', updateCluster_replicaConfiguration - The number of replicas that will reside in each shard

$sel:nodeType:UpdateCluster', updateCluster_nodeType - A valid node type that you want to scale this cluster up or down to.

$sel:snsTopicStatus:UpdateCluster', updateCluster_snsTopicStatus - The status of the Amazon SNS notification topic. Notifications are sent only if the status is active.

$sel:description:UpdateCluster', updateCluster_description - The description of the cluster to update

$sel:aCLName:UpdateCluster', updateCluster_aCLName - The Access Control List that is associated with the cluster

$sel:parameterGroupName:UpdateCluster', updateCluster_parameterGroupName - The name of the parameter group to update

$sel:clusterName:UpdateCluster', updateCluster_clusterName - The name of the cluster to update

Request Lenses

updateCluster_engineVersion :: Lens' UpdateCluster (Maybe Text)

The upgraded version of the engine to be run on the nodes. You can upgrade to a newer engine version, but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cluster and create it anew with the earlier engine version.

updateCluster_securityGroupIds :: Lens' UpdateCluster (Maybe [Text])

The SecurityGroupIds to update

updateCluster_snsTopicArn :: Lens' UpdateCluster (Maybe Text)

The SNS topic ARN to update

updateCluster_snapshotWindow :: Lens' UpdateCluster (Maybe Text)

The daily time range (in UTC) during which MemoryDB begins taking a daily snapshot of your cluster.

updateCluster_maintenanceWindow :: Lens' UpdateCluster (Maybe Text)

The maintenance window to update

updateCluster_shardConfiguration :: Lens' UpdateCluster (Maybe ShardConfigurationRequest)

The number of shards in the cluster

updateCluster_snapshotRetentionLimit :: Lens' UpdateCluster (Maybe Int)

The number of days for which MemoryDB retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

updateCluster_replicaConfiguration :: Lens' UpdateCluster (Maybe ReplicaConfigurationRequest)

The number of replicas that will reside in each shard

updateCluster_nodeType :: Lens' UpdateCluster (Maybe Text)

A valid node type that you want to scale this cluster up or down to.

updateCluster_snsTopicStatus :: Lens' UpdateCluster (Maybe Text)

The status of the Amazon SNS notification topic. Notifications are sent only if the status is active.

updateCluster_description :: Lens' UpdateCluster (Maybe Text)

The description of the cluster to update

updateCluster_aCLName :: Lens' UpdateCluster (Maybe Text)

The Access Control List that is associated with the cluster

updateCluster_parameterGroupName :: Lens' UpdateCluster (Maybe Text)

The name of the parameter group to update

updateCluster_clusterName :: Lens' UpdateCluster Text

The name of the cluster to update

Destructuring the Response

data UpdateClusterResponse

See: newUpdateClusterResponse smart constructor.

Constructors

UpdateClusterResponse'

Fields

cluster :: Maybe Cluster

The updated cluster

httpStatus :: Int

The response's http status code.

newUpdateClusterResponse

Arguments

:: Int
$sel:httpStatus:UpdateClusterResponse'

-> UpdateClusterResponse

Create a value of UpdateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:UpdateClusterResponse', updateClusterResponse_cluster - The updated cluster

$sel:httpStatus:UpdateClusterResponse', updateClusterResponse_httpStatus - The response's http status code.

Response Lenses

updateClusterResponse_cluster :: Lens' UpdateClusterResponse (Maybe Cluster)

The updated cluster

updateClusterResponse_httpStatus :: Lens' UpdateClusterResponse Int

The response's http status code.