Amazonka.MigrationHub.AssociateCreatedArtifact

Description

Associates a created artifact of an AWS cloud resource, the target receiving the migration, with the migration task performed by a migration tool. This API has the following traits:

Migration tools can call the AssociateCreatedArtifact operation to indicate which AWS artifact is associated with a migration task.

The created artifact name must be provided in ARN (Amazon Resource Name) format which will contain information about type and region; for example: arn:aws:ec2:us-east-1:488216288981:image/ami-6d0ba87b.

Examples of the AWS resource behind the created artifact are, AMI's, EC2 instance, or DMS endpoint, etc.

Creating a Request

data AssociateCreatedArtifact

See: newAssociateCreatedArtifact smart constructor.

Constructors

AssociateCreatedArtifact'

Fields

dryRun :: Maybe Bool

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

progressUpdateStream :: Text

The name of the ProgressUpdateStream.

migrationTaskName :: Text

Unique identifier that references the migration task. /Do not store personal data in this field./

createdArtifact :: CreatedArtifact

An ARN of the AWS resource related to the migration (e.g., AMI, EC2 instance, RDS instance, etc.)

newAssociateCreatedArtifact

Arguments

:: Text
$sel:progressUpdateStream:AssociateCreatedArtifact'

-> Text
$sel:migrationTaskName:AssociateCreatedArtifact'

-> CreatedArtifact
$sel:createdArtifact:AssociateCreatedArtifact'

-> AssociateCreatedArtifact

Create a value of AssociateCreatedArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:AssociateCreatedArtifact', associateCreatedArtifact_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStream:AssociateCreatedArtifact', associateCreatedArtifact_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:AssociateCreatedArtifact', associateCreatedArtifact_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

$sel:createdArtifact:AssociateCreatedArtifact', associateCreatedArtifact_createdArtifact - An ARN of the AWS resource related to the migration (e.g., AMI, EC2 instance, RDS instance, etc.)

Request Lenses

associateCreatedArtifact_dryRun :: Lens' AssociateCreatedArtifact (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

associateCreatedArtifact_progressUpdateStream :: Lens' AssociateCreatedArtifact Text

The name of the ProgressUpdateStream.

associateCreatedArtifact_migrationTaskName :: Lens' AssociateCreatedArtifact Text

Unique identifier that references the migration task. /Do not store personal data in this field./

associateCreatedArtifact_createdArtifact :: Lens' AssociateCreatedArtifact CreatedArtifact

An ARN of the AWS resource related to the migration (e.g., AMI, EC2 instance, RDS instance, etc.)

Destructuring the Response

data AssociateCreatedArtifactResponse

See: newAssociateCreatedArtifactResponse smart constructor.

Constructors

AssociateCreatedArtifactResponse'

Fields

httpStatus :: Int

The response's http status code.

newAssociateCreatedArtifactResponse

Arguments

:: Int
$sel:httpStatus:AssociateCreatedArtifactResponse'

-> AssociateCreatedArtifactResponse

Create a value of AssociateCreatedArtifactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateCreatedArtifactResponse', associateCreatedArtifactResponse_httpStatus - The response's http status code.

Response Lenses

associateCreatedArtifactResponse_httpStatus :: Lens' AssociateCreatedArtifactResponse Int

The response's http status code.