Amazonka.MigrationHub.DisassociateDiscoveredResource

Description

Disassociate an Application Discovery Service discovered resource from a migration task.

Creating a Request

data DisassociateDiscoveredResource

See: newDisassociateDiscoveredResource smart constructor.

Constructors

DisassociateDiscoveredResource'

Fields

dryRun :: Maybe Bool

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

progressUpdateStream :: Text

The name of the ProgressUpdateStream.

migrationTaskName :: Text

The identifier given to the MigrationTask. /Do not store personal data in this field./

configurationId :: Text

ConfigurationId of the Application Discovery Service resource to be disassociated.

newDisassociateDiscoveredResource

Arguments

:: Text
$sel:progressUpdateStream:DisassociateDiscoveredResource'

-> Text
$sel:migrationTaskName:DisassociateDiscoveredResource'

-> Text
$sel:configurationId:DisassociateDiscoveredResource'

-> DisassociateDiscoveredResource

Create a value of DisassociateDiscoveredResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DisassociateDiscoveredResource', disassociateDiscoveredResource_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStream:DisassociateDiscoveredResource', disassociateDiscoveredResource_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:DisassociateDiscoveredResource', disassociateDiscoveredResource_migrationTaskName - The identifier given to the MigrationTask. /Do not store personal data in this field./

$sel:configurationId:DisassociateDiscoveredResource', disassociateDiscoveredResource_configurationId - ConfigurationId of the Application Discovery Service resource to be disassociated.

Request Lenses

disassociateDiscoveredResource_dryRun :: Lens' DisassociateDiscoveredResource (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

disassociateDiscoveredResource_progressUpdateStream :: Lens' DisassociateDiscoveredResource Text

The name of the ProgressUpdateStream.

disassociateDiscoveredResource_migrationTaskName :: Lens' DisassociateDiscoveredResource Text

The identifier given to the MigrationTask. /Do not store personal data in this field./

disassociateDiscoveredResource_configurationId :: Lens' DisassociateDiscoveredResource Text

ConfigurationId of the Application Discovery Service resource to be disassociated.

Destructuring the Response

data DisassociateDiscoveredResourceResponse

See: newDisassociateDiscoveredResourceResponse smart constructor.

Constructors

DisassociateDiscoveredResourceResponse'

Fields

httpStatus :: Int

The response's http status code.

newDisassociateDiscoveredResourceResponse

Arguments

:: Int
$sel:httpStatus:DisassociateDiscoveredResourceResponse'

-> DisassociateDiscoveredResourceResponse

Create a value of DisassociateDiscoveredResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateDiscoveredResourceResponse', disassociateDiscoveredResourceResponse_httpStatus - The response's http status code.

Response Lenses

disassociateDiscoveredResourceResponse_httpStatus :: Lens' DisassociateDiscoveredResourceResponse Int

The response's http status code.