Amazonka.MigrationHub.Lens

Description

Operations

AssociateDiscoveredResource

associateDiscoveredResource_dryRun :: Lens' AssociateDiscoveredResource (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

associateDiscoveredResource_progressUpdateStream :: Lens' AssociateDiscoveredResource Text

The name of the ProgressUpdateStream.

associateDiscoveredResource_migrationTaskName :: Lens' AssociateDiscoveredResource Text

The identifier given to the MigrationTask. /Do not store personal data in this field./

associateDiscoveredResource_discoveredResource :: Lens' AssociateDiscoveredResource DiscoveredResource

Object representing a Resource.

associateDiscoveredResourceResponse_httpStatus :: Lens' AssociateDiscoveredResourceResponse Int

The response's http status code.

CreateProgressUpdateStream

createProgressUpdateStream_dryRun :: Lens' CreateProgressUpdateStream (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

createProgressUpdateStream_progressUpdateStreamName :: Lens' CreateProgressUpdateStream Text

The name of the ProgressUpdateStream. /Do not store personal data in this field./

createProgressUpdateStreamResponse_httpStatus :: Lens' CreateProgressUpdateStreamResponse Int

The response's http status code.

ListApplicationStates

listApplicationStates_nextToken :: Lens' ListApplicationStates (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

listApplicationStates_applicationIds :: Lens' ListApplicationStates (Maybe (NonEmpty Text))

The configurationIds from the Application Discovery Service that uniquely identifies your applications.

listApplicationStates_maxResults :: Lens' ListApplicationStates (Maybe Natural)

Maximum number of results to be returned per page.

listApplicationStatesResponse_applicationStateList :: Lens' ListApplicationStatesResponse (Maybe [ApplicationState])

A list of Applications that exist in Application Discovery Service.

listApplicationStatesResponse_nextToken :: Lens' ListApplicationStatesResponse (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

listApplicationStatesResponse_httpStatus :: Lens' ListApplicationStatesResponse Int

The response's http status code.

ListCreatedArtifacts

listCreatedArtifacts_nextToken :: Lens' ListCreatedArtifacts (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

listCreatedArtifacts_maxResults :: Lens' ListCreatedArtifacts (Maybe Natural)

Maximum number of results to be returned per page.

listCreatedArtifacts_progressUpdateStream :: Lens' ListCreatedArtifacts Text

The name of the ProgressUpdateStream.

listCreatedArtifacts_migrationTaskName :: Lens' ListCreatedArtifacts Text

Unique identifier that references the migration task. /Do not store personal data in this field./

listCreatedArtifactsResponse_nextToken :: Lens' ListCreatedArtifactsResponse (Maybe Text)

If there are more created artifacts than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

listCreatedArtifactsResponse_createdArtifactList :: Lens' ListCreatedArtifactsResponse (Maybe [CreatedArtifact])

List of created artifacts up to the maximum number of results specified in the request.

listCreatedArtifactsResponse_httpStatus :: Lens' ListCreatedArtifactsResponse Int

The response's http status code.

DisassociateDiscoveredResource

disassociateDiscoveredResource_dryRun :: Lens' DisassociateDiscoveredResource (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

disassociateDiscoveredResource_progressUpdateStream :: Lens' DisassociateDiscoveredResource Text

The name of the ProgressUpdateStream.

disassociateDiscoveredResource_migrationTaskName :: Lens' DisassociateDiscoveredResource Text

The identifier given to the MigrationTask. /Do not store personal data in this field./

disassociateDiscoveredResource_configurationId :: Lens' DisassociateDiscoveredResource Text

ConfigurationId of the Application Discovery Service resource to be disassociated.

disassociateDiscoveredResourceResponse_httpStatus :: Lens' DisassociateDiscoveredResourceResponse Int

The response's http status code.

NotifyApplicationState

notifyApplicationState_updateDateTime :: Lens' NotifyApplicationState (Maybe UTCTime)

The timestamp when the application state changed.

notifyApplicationState_dryRun :: Lens' NotifyApplicationState (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

notifyApplicationState_applicationId :: Lens' NotifyApplicationState Text

The configurationId in Application Discovery Service that uniquely identifies the grouped application.

notifyApplicationState_status :: Lens' NotifyApplicationState ApplicationStatus

Status of the application - Not Started, In-Progress, Complete.

notifyApplicationStateResponse_httpStatus :: Lens' NotifyApplicationStateResponse Int

The response's http status code.

DeleteProgressUpdateStream

deleteProgressUpdateStream_dryRun :: Lens' DeleteProgressUpdateStream (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

deleteProgressUpdateStream_progressUpdateStreamName :: Lens' DeleteProgressUpdateStream Text

The name of the ProgressUpdateStream. /Do not store personal data in this field./

deleteProgressUpdateStreamResponse_httpStatus :: Lens' DeleteProgressUpdateStreamResponse Int

The response's http status code.

ListProgressUpdateStreams

listProgressUpdateStreams_nextToken :: Lens' ListProgressUpdateStreams (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

listProgressUpdateStreams_maxResults :: Lens' ListProgressUpdateStreams (Maybe Natural)

Filter to limit the maximum number of results to list per page.

listProgressUpdateStreamsResponse_progressUpdateStreamSummaryList :: Lens' ListProgressUpdateStreamsResponse (Maybe [ProgressUpdateStreamSummary])

List of progress update streams up to the max number of results passed in the input.

listProgressUpdateStreamsResponse_nextToken :: Lens' ListProgressUpdateStreamsResponse (Maybe Text)

If there are more streams created than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

listProgressUpdateStreamsResponse_httpStatus :: Lens' ListProgressUpdateStreamsResponse Int

The response's http status code.

DisassociateCreatedArtifact

disassociateCreatedArtifact_dryRun :: Lens' DisassociateCreatedArtifact (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

disassociateCreatedArtifact_progressUpdateStream :: Lens' DisassociateCreatedArtifact Text

The name of the ProgressUpdateStream.

disassociateCreatedArtifact_migrationTaskName :: Lens' DisassociateCreatedArtifact Text

Unique identifier that references the migration task to be disassociated with the artifact. Do not store personal data in this field.

disassociateCreatedArtifact_createdArtifactName :: Lens' DisassociateCreatedArtifact Text

An ARN of the AWS resource related to the migration (e.g., AMI, EC2 instance, RDS instance, etc.)

disassociateCreatedArtifactResponse_httpStatus :: Lens' DisassociateCreatedArtifactResponse Int

The response's http status code.

ImportMigrationTask

importMigrationTask_dryRun :: Lens' ImportMigrationTask (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

importMigrationTask_progressUpdateStream :: Lens' ImportMigrationTask Text

The name of the ProgressUpdateStream. >

importMigrationTask_migrationTaskName :: Lens' ImportMigrationTask Text

Unique identifier that references the migration task. /Do not store personal data in this field./

importMigrationTaskResponse_httpStatus :: Lens' ImportMigrationTaskResponse Int

The response's http status code.

DescribeMigrationTask

describeMigrationTask_progressUpdateStream :: Lens' DescribeMigrationTask Text

The name of the ProgressUpdateStream.

describeMigrationTask_migrationTaskName :: Lens' DescribeMigrationTask Text

The identifier given to the MigrationTask. /Do not store personal data in this field./

describeMigrationTaskResponse_migrationTask :: Lens' DescribeMigrationTaskResponse (Maybe MigrationTask)

Object encapsulating information about the migration task.

describeMigrationTaskResponse_httpStatus :: Lens' DescribeMigrationTaskResponse Int

The response's http status code.

PutResourceAttributes

putResourceAttributes_dryRun :: Lens' PutResourceAttributes (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

putResourceAttributes_progressUpdateStream :: Lens' PutResourceAttributes Text

The name of the ProgressUpdateStream.

putResourceAttributes_migrationTaskName :: Lens' PutResourceAttributes Text

Unique identifier that references the migration task. /Do not store personal data in this field./

putResourceAttributes_resourceAttributeList :: Lens' PutResourceAttributes (NonEmpty ResourceAttribute)

Information about the resource that is being migrated. This data will be used to map the task to a resource in the Application Discovery Service repository.

Takes the object array of ResourceAttribute where the Type field is reserved for the following values: IPV4_ADDRESS | IPV6_ADDRESS | MAC_ADDRESS | FQDN | VM_MANAGER_ID | VM_MANAGED_OBJECT_REFERENCE | VM_NAME | VM_PATH | BIOS_ID | MOTHERBOARD_SERIAL_NUMBER where the identifying value can be a string up to 256 characters.

If any "VM" related value is set for a ResourceAttribute object, it is required that VM_MANAGER_ID, as a minimum, is always set. If VM_MANAGER_ID is not set, then all "VM" fields will be discarded and "VM" fields will not be used for matching the migration task to a server in Application Discovery Service repository. See the Example section below for a use case of specifying "VM" related values.

If a server you are trying to match has multiple IP or MAC addresses, you should provide as many as you know in separate type/value pairs passed to the ResourceAttributeList parameter to maximize the chances of matching.

putResourceAttributesResponse_httpStatus :: Lens' PutResourceAttributesResponse Int

The response's http status code.

AssociateCreatedArtifact

associateCreatedArtifact_dryRun :: Lens' AssociateCreatedArtifact (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

associateCreatedArtifact_progressUpdateStream :: Lens' AssociateCreatedArtifact Text

The name of the ProgressUpdateStream.

associateCreatedArtifact_migrationTaskName :: Lens' AssociateCreatedArtifact Text

Unique identifier that references the migration task. /Do not store personal data in this field./

associateCreatedArtifact_createdArtifact :: Lens' AssociateCreatedArtifact CreatedArtifact

An ARN of the AWS resource related to the migration (e.g., AMI, EC2 instance, RDS instance, etc.)

associateCreatedArtifactResponse_httpStatus :: Lens' AssociateCreatedArtifactResponse Int

The response's http status code.

NotifyMigrationTaskState

notifyMigrationTaskState_dryRun :: Lens' NotifyMigrationTaskState (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

notifyMigrationTaskState_progressUpdateStream :: Lens' NotifyMigrationTaskState Text

The name of the ProgressUpdateStream.

notifyMigrationTaskState_migrationTaskName :: Lens' NotifyMigrationTaskState Text

Unique identifier that references the migration task. /Do not store personal data in this field./

notifyMigrationTaskState_task :: Lens' NotifyMigrationTaskState Task

Information about the task's progress and status.

notifyMigrationTaskState_updateDateTime :: Lens' NotifyMigrationTaskState UTCTime

The timestamp when the task was gathered.

notifyMigrationTaskState_nextUpdateSeconds :: Lens' NotifyMigrationTaskState Natural

Number of seconds after the UpdateDateTime within which the Migration Hub can expect an update. If Migration Hub does not receive an update within the specified interval, then the migration task will be considered stale.

notifyMigrationTaskStateResponse_httpStatus :: Lens' NotifyMigrationTaskStateResponse Int

The response's http status code.

DescribeApplicationState

describeApplicationState_applicationId :: Lens' DescribeApplicationState Text

The configurationId in Application Discovery Service that uniquely identifies the grouped application.

describeApplicationStateResponse_lastUpdatedTime :: Lens' DescribeApplicationStateResponse (Maybe UTCTime)

The timestamp when the application status was last updated.

describeApplicationStateResponse_applicationStatus :: Lens' DescribeApplicationStateResponse (Maybe ApplicationStatus)

Status of the application - Not Started, In-Progress, Complete.

describeApplicationStateResponse_httpStatus :: Lens' DescribeApplicationStateResponse Int

The response's http status code.

ListMigrationTasks

listMigrationTasks_resourceName :: Lens' ListMigrationTasks (Maybe Text)

Filter migration tasks by discovered resource name.

listMigrationTasks_nextToken :: Lens' ListMigrationTasks (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

listMigrationTasks_maxResults :: Lens' ListMigrationTasks (Maybe Natural)

Value to specify how many results are returned per page.

listMigrationTasksResponse_migrationTaskSummaryList :: Lens' ListMigrationTasksResponse (Maybe [MigrationTaskSummary])

Lists the migration task's summary which includes: MigrationTaskName, ProgressPercent, ProgressUpdateStream, Status, and the UpdateDateTime for each task.

listMigrationTasksResponse_nextToken :: Lens' ListMigrationTasksResponse (Maybe Text)

If there are more migration tasks than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

listMigrationTasksResponse_httpStatus :: Lens' ListMigrationTasksResponse Int

The response's http status code.

ListDiscoveredResources

listDiscoveredResources_nextToken :: Lens' ListDiscoveredResources (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

listDiscoveredResources_maxResults :: Lens' ListDiscoveredResources (Maybe Natural)

The maximum number of results returned per page.

listDiscoveredResources_progressUpdateStream :: Lens' ListDiscoveredResources Text

The name of the ProgressUpdateStream.

listDiscoveredResources_migrationTaskName :: Lens' ListDiscoveredResources Text

The name of the MigrationTask. /Do not store personal data in this field./

listDiscoveredResourcesResponse_discoveredResourceList :: Lens' ListDiscoveredResourcesResponse (Maybe [DiscoveredResource])

Returned list of discovered resources associated with the given MigrationTask.

listDiscoveredResourcesResponse_nextToken :: Lens' ListDiscoveredResourcesResponse (Maybe Text)

If there are more discovered resources than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

listDiscoveredResourcesResponse_httpStatus :: Lens' ListDiscoveredResourcesResponse Int

The response's http status code.

Types

ApplicationState

applicationState_lastUpdatedTime :: Lens' ApplicationState (Maybe UTCTime)

The timestamp when the application status was last updated.

applicationState_applicationId :: Lens' ApplicationState (Maybe Text)

The configurationId from the Application Discovery Service that uniquely identifies an application.

applicationState_applicationStatus :: Lens' ApplicationState (Maybe ApplicationStatus)

The current status of an application.

CreatedArtifact

createdArtifact_description :: Lens' CreatedArtifact (Maybe Text)

A description that can be free-form text to record additional detail about the artifact for clarity or for later reference.

createdArtifact_name :: Lens' CreatedArtifact Text

An ARN that uniquely identifies the result of a migration task.

DiscoveredResource

discoveredResource_description :: Lens' DiscoveredResource (Maybe Text)

A description that can be free-form text to record additional detail about the discovered resource for clarity or later reference.

discoveredResource_configurationId :: Lens' DiscoveredResource Text

The configurationId in Application Discovery Service that uniquely identifies the on-premise resource.

MigrationTask

migrationTask_updateDateTime :: Lens' MigrationTask (Maybe UTCTime)

The timestamp when the task was gathered.

migrationTask_resourceAttributeList :: Lens' MigrationTask (Maybe [ResourceAttribute])

Information about the resource that is being migrated. This data will be used to map the task to a resource in the Application Discovery Service repository.

migrationTask_task :: Lens' MigrationTask (Maybe Task)

Task object encapsulating task information.

migrationTask_progressUpdateStream :: Lens' MigrationTask (Maybe Text)

A name that identifies the vendor of the migration tool being used.

migrationTask_migrationTaskName :: Lens' MigrationTask (Maybe Text)

Unique identifier that references the migration task. /Do not store personal data in this field./

MigrationTaskSummary

migrationTaskSummary_status :: Lens' MigrationTaskSummary (Maybe MigrationStatus)

Status of the task.

migrationTaskSummary_updateDateTime :: Lens' MigrationTaskSummary (Maybe UTCTime)

The timestamp when the task was gathered.

migrationTaskSummary_progressPercent :: Lens' MigrationTaskSummary (Maybe Natural)

Indication of the percentage completion of the task.

migrationTaskSummary_statusDetail :: Lens' MigrationTaskSummary (Maybe Text)

Detail information of what is being done within the overall status state.

migrationTaskSummary_progressUpdateStream :: Lens' MigrationTaskSummary (Maybe Text)

An AWS resource used for access control. It should uniquely identify the migration tool as it is used for all updates made by the tool.

migrationTaskSummary_migrationTaskName :: Lens' MigrationTaskSummary (Maybe Text)

Unique identifier that references the migration task. /Do not store personal data in this field./

ProgressUpdateStreamSummary

progressUpdateStreamSummary_progressUpdateStreamName :: Lens' ProgressUpdateStreamSummary (Maybe Text)

The name of the ProgressUpdateStream. /Do not store personal data in this field./

ResourceAttribute

resourceAttribute_type :: Lens' ResourceAttribute ResourceAttributeType

Type of resource.

resourceAttribute_value :: Lens' ResourceAttribute Text

Value of the resource type.

Task

task_progressPercent :: Lens' Task (Maybe Natural)

Indication of the percentage completion of the task.

task_statusDetail :: Lens' Task (Maybe Text)

Details of task status as notified by a migration tool. A tool might use this field to provide clarifying information about the status that is unique to that tool or that explains an error state.

task_status :: Lens' Task MigrationStatus

Status of the task - Not Started, In-Progress, Complete.