Amazonka.MigrationHubConfig

Description

Derived from API version 2019-06-30 of the AWS service descriptions, licensed under Apache 2.0.

The AWS Migration Hub home region APIs are available specifically for working with your Migration Hub home region. You can use these APIs to determine a home region, as well as to create and work with controls that describe the home region.

You must make API calls for write actions (create, notify, associate, disassociate, import, or put) while in your home region, or a HomeRegionNotSetException error is returned.

API calls for read actions (list, describe, stop, and delete) are permitted outside of your home region.

If you call a write API outside the home region, an InvalidInputException is returned.

You can call GetHomeRegion action to obtain the account's Migration Hub home region.

For specific API usage, see the sections that follow in this AWS Migration Hub Home Region API reference.

Service Configuration

defaultService :: Service

API version 2019-06-30 of the Amazon Migration Hub Config SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MigrationHubConfig.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

DryRunOperation

_DryRunOperation :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised to indicate that authorization of an action was successful, when the DryRun flag is set to true.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

InternalServerError

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when an internal, configuration, or dependency error is encountered.

InvalidInputException

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when the provided input violates a policy constraint or is entered in the wrong format or data type.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when a request fails due to temporary unavailability of the service.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetHomeRegion

data GetHomeRegion

See: newGetHomeRegion smart constructor.

Constructors

GetHomeRegion'

newGetHomeRegion :: GetHomeRegion

Create a value of GetHomeRegion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

data GetHomeRegionResponse

See: newGetHomeRegionResponse smart constructor.

Constructors

GetHomeRegionResponse' (Maybe Text) Int

newGetHomeRegionResponse

Arguments

:: Int
$sel:httpStatus:GetHomeRegionResponse'

-> GetHomeRegionResponse

Create a value of GetHomeRegionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:homeRegion:GetHomeRegionResponse', getHomeRegionResponse_homeRegion - The name of the home region of the calling account.

$sel:httpStatus:GetHomeRegionResponse', getHomeRegionResponse_httpStatus - The response's http status code.

CreateHomeRegionControl

data CreateHomeRegionControl

See: newCreateHomeRegionControl smart constructor.

Constructors

CreateHomeRegionControl' (Maybe Bool) Text Target

newCreateHomeRegionControl

Arguments

:: Text
$sel:homeRegion:CreateHomeRegionControl'

-> Target
$sel:target:CreateHomeRegionControl'

-> CreateHomeRegionControl

Create a value of CreateHomeRegionControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:CreateHomeRegionControl', createHomeRegionControl_dryRun - Optional Boolean flag to indicate whether any effect should take place. It tests whether the caller has permission to make the call.

$sel:homeRegion:CreateHomeRegionControl', createHomeRegionControl_homeRegion - The name of the home region of the calling account.

$sel:target:CreateHomeRegionControl', createHomeRegionControl_target - The account for which this command sets up a home region control. The Target is always of type ACCOUNT.

data CreateHomeRegionControlResponse

See: newCreateHomeRegionControlResponse smart constructor.

Constructors

CreateHomeRegionControlResponse' (Maybe HomeRegionControl) Int

newCreateHomeRegionControlResponse

Arguments

:: Int
$sel:httpStatus:CreateHomeRegionControlResponse'

-> CreateHomeRegionControlResponse

Create a value of CreateHomeRegionControlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:homeRegionControl:CreateHomeRegionControlResponse', createHomeRegionControlResponse_homeRegionControl - This object is the HomeRegionControl object that's returned by a successful call to CreateHomeRegionControl.

$sel:httpStatus:CreateHomeRegionControlResponse', createHomeRegionControlResponse_httpStatus - The response's http status code.

DescribeHomeRegionControls

data DescribeHomeRegionControls

See: newDescribeHomeRegionControls smart constructor.

Constructors

DescribeHomeRegionControls' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Natural) (Maybe Target)

newDescribeHomeRegionControls :: DescribeHomeRegionControls

Create a value of DescribeHomeRegionControls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeHomeRegionControls', describeHomeRegionControls_nextToken - If a NextToken was returned by a previous call, more results are available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:homeRegion:DescribeHomeRegionControls', describeHomeRegionControls_homeRegion - The name of the home region you'd like to view.

$sel:controlId:DescribeHomeRegionControls', describeHomeRegionControls_controlId - The ControlID is a unique identifier string of your HomeRegionControl object.

$sel:maxResults:DescribeHomeRegionControls', describeHomeRegionControls_maxResults - The maximum number of filtering results to display per page.

$sel:target:DescribeHomeRegionControls', describeHomeRegionControls_target - The target parameter specifies the identifier to which the home region is applied, which is always of type ACCOUNT. It applies the home region to the current ACCOUNT.

data DescribeHomeRegionControlsResponse

See: newDescribeHomeRegionControlsResponse smart constructor.

Constructors

DescribeHomeRegionControlsResponse' (Maybe Text) (Maybe [HomeRegionControl]) Int

newDescribeHomeRegionControlsResponse

Arguments

:: Int
$sel:httpStatus:DescribeHomeRegionControlsResponse'

-> DescribeHomeRegionControlsResponse

Create a value of DescribeHomeRegionControlsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeHomeRegionControls', describeHomeRegionControlsResponse_nextToken - If a NextToken was returned by a previous call, more results are available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:homeRegionControls:DescribeHomeRegionControlsResponse', describeHomeRegionControlsResponse_homeRegionControls - An array that contains your HomeRegionControl objects.

$sel:httpStatus:DescribeHomeRegionControlsResponse', describeHomeRegionControlsResponse_httpStatus - The response's http status code.

Types

TargetType

newtype TargetType

Constructors

TargetType'

Fields

fromTargetType :: Text

Bundled Patterns

pattern TargetType_ACCOUNT :: TargetType

HomeRegionControl

data HomeRegionControl

A home region control is an object that specifies the home region for an account, with some additional information. It contains a target (always of type ACCOUNT), an ID, and a time at which the home region was set.

See: newHomeRegionControl smart constructor.

Constructors

HomeRegionControl' (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Target)

newHomeRegionControl :: HomeRegionControl

Create a value of HomeRegionControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestedTime:HomeRegionControl', homeRegionControl_requestedTime - A timestamp representing the time when the customer called CreateHomeregionControl and set the home region for the account.

$sel:homeRegion:HomeRegionControl', homeRegionControl_homeRegion - The AWS Region that's been set as home region. For example, "us-west-2" or "eu-central-1" are valid home regions.

$sel:controlId:HomeRegionControl', homeRegionControl_controlId - A unique identifier that's generated for each home region control. It's always a string that begins with "hrc-" followed by 12 lowercase letters and numbers.

$sel:target:HomeRegionControl', homeRegionControl_target - The target parameter specifies the identifier to which the home region is applied, which is always an ACCOUNT. It applies the home region to the current ACCOUNT.

Target

data Target

The target parameter specifies the identifier to which the home region is applied, which is always an ACCOUNT. It applies the home region to the current ACCOUNT.

See: newTarget smart constructor.

Constructors

Target' (Maybe Text) TargetType

newTarget

Arguments

:: TargetType
$sel:type':Target'

-> Target

Create a value of Target with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Target', target_id - The TargetID is a 12-character identifier of the ACCOUNT for which the control was created. (This must be the current account.)

$sel:type':Target', target_type - The target type is always an ACCOUNT.