Amazonka.MigrationHubConfig.CreateHomeRegionControl

Description

This API sets up the home region for the calling account only.

Creating a Request

data CreateHomeRegionControl

See: newCreateHomeRegionControl smart constructor.

Constructors

CreateHomeRegionControl'

Fields

dryRun :: Maybe Bool

Optional Boolean flag to indicate whether any effect should take place. It tests whether the caller has permission to make the call.

homeRegion :: Text

The name of the home region of the calling account.

target :: Target

The account for which this command sets up a home region control. The Target is always of type ACCOUNT.

newCreateHomeRegionControl

Arguments

:: Text
$sel:homeRegion:CreateHomeRegionControl'

-> Target
$sel:target:CreateHomeRegionControl'

-> CreateHomeRegionControl

Create a value of CreateHomeRegionControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:CreateHomeRegionControl', createHomeRegionControl_dryRun - Optional Boolean flag to indicate whether any effect should take place. It tests whether the caller has permission to make the call.

$sel:homeRegion:CreateHomeRegionControl', createHomeRegionControl_homeRegion - The name of the home region of the calling account.

$sel:target:CreateHomeRegionControl', createHomeRegionControl_target - The account for which this command sets up a home region control. The Target is always of type ACCOUNT.

Request Lenses

createHomeRegionControl_dryRun :: Lens' CreateHomeRegionControl (Maybe Bool)

Optional Boolean flag to indicate whether any effect should take place. It tests whether the caller has permission to make the call.

createHomeRegionControl_homeRegion :: Lens' CreateHomeRegionControl Text

The name of the home region of the calling account.

createHomeRegionControl_target :: Lens' CreateHomeRegionControl Target

The account for which this command sets up a home region control. The Target is always of type ACCOUNT.

Destructuring the Response

data CreateHomeRegionControlResponse

See: newCreateHomeRegionControlResponse smart constructor.

Constructors

CreateHomeRegionControlResponse'

Fields

homeRegionControl :: Maybe HomeRegionControl

This object is the HomeRegionControl object that's returned by a successful call to CreateHomeRegionControl.

httpStatus :: Int

The response's http status code.

newCreateHomeRegionControlResponse

Arguments

:: Int
$sel:httpStatus:CreateHomeRegionControlResponse'

-> CreateHomeRegionControlResponse

Create a value of CreateHomeRegionControlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:homeRegionControl:CreateHomeRegionControlResponse', createHomeRegionControlResponse_homeRegionControl - This object is the HomeRegionControl object that's returned by a successful call to CreateHomeRegionControl.

$sel:httpStatus:CreateHomeRegionControlResponse', createHomeRegionControlResponse_httpStatus - The response's http status code.

Response Lenses

createHomeRegionControlResponse_homeRegionControl :: Lens' CreateHomeRegionControlResponse (Maybe HomeRegionControl)

This object is the HomeRegionControl object that's returned by a successful call to CreateHomeRegionControl.

createHomeRegionControlResponse_httpStatus :: Lens' CreateHomeRegionControlResponse Int

The response's http status code.