Amazonka.MigrationHubConfig.DescribeHomeRegionControls

Description

This API permits filtering on the ControlId and HomeRegion fields.

Creating a Request

data DescribeHomeRegionControls

See: newDescribeHomeRegionControls smart constructor.

Constructors

DescribeHomeRegionControls'

Fields

nextToken :: Maybe Text

If a NextToken was returned by a previous call, more results are available. To retrieve the next page of results, make the call again using the returned token in NextToken.

homeRegion :: Maybe Text

The name of the home region you'd like to view.

controlId :: Maybe Text

The ControlID is a unique identifier string of your HomeRegionControl object.

maxResults :: Maybe Natural

The maximum number of filtering results to display per page.

target :: Maybe Target

The target parameter specifies the identifier to which the home region is applied, which is always of type ACCOUNT. It applies the home region to the current ACCOUNT.

newDescribeHomeRegionControls :: DescribeHomeRegionControls

Create a value of DescribeHomeRegionControls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeHomeRegionControls', describeHomeRegionControls_nextToken - If a NextToken was returned by a previous call, more results are available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:homeRegion:DescribeHomeRegionControls', describeHomeRegionControls_homeRegion - The name of the home region you'd like to view.

$sel:controlId:DescribeHomeRegionControls', describeHomeRegionControls_controlId - The ControlID is a unique identifier string of your HomeRegionControl object.

$sel:maxResults:DescribeHomeRegionControls', describeHomeRegionControls_maxResults - The maximum number of filtering results to display per page.

$sel:target:DescribeHomeRegionControls', describeHomeRegionControls_target - The target parameter specifies the identifier to which the home region is applied, which is always of type ACCOUNT. It applies the home region to the current ACCOUNT.

Request Lenses

describeHomeRegionControls_nextToken :: Lens' DescribeHomeRegionControls (Maybe Text)

If a NextToken was returned by a previous call, more results are available. To retrieve the next page of results, make the call again using the returned token in NextToken.

describeHomeRegionControls_homeRegion :: Lens' DescribeHomeRegionControls (Maybe Text)

The name of the home region you'd like to view.

describeHomeRegionControls_controlId :: Lens' DescribeHomeRegionControls (Maybe Text)

The ControlID is a unique identifier string of your HomeRegionControl object.

describeHomeRegionControls_maxResults :: Lens' DescribeHomeRegionControls (Maybe Natural)

The maximum number of filtering results to display per page.

describeHomeRegionControls_target :: Lens' DescribeHomeRegionControls (Maybe Target)

The target parameter specifies the identifier to which the home region is applied, which is always of type ACCOUNT. It applies the home region to the current ACCOUNT.

Destructuring the Response

data DescribeHomeRegionControlsResponse

See: newDescribeHomeRegionControlsResponse smart constructor.

Constructors

DescribeHomeRegionControlsResponse'

Fields

nextToken :: Maybe Text

If a NextToken was returned by a previous call, more results are available. To retrieve the next page of results, make the call again using the returned token in NextToken.

homeRegionControls :: Maybe [HomeRegionControl]

An array that contains your HomeRegionControl objects.

httpStatus :: Int

The response's http status code.

newDescribeHomeRegionControlsResponse

Arguments

:: Int
$sel:httpStatus:DescribeHomeRegionControlsResponse'

-> DescribeHomeRegionControlsResponse

Create a value of DescribeHomeRegionControlsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeHomeRegionControls', describeHomeRegionControlsResponse_nextToken - If a NextToken was returned by a previous call, more results are available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:homeRegionControls:DescribeHomeRegionControlsResponse', describeHomeRegionControlsResponse_homeRegionControls - An array that contains your HomeRegionControl objects.

$sel:httpStatus:DescribeHomeRegionControlsResponse', describeHomeRegionControlsResponse_httpStatus - The response's http status code.

Response Lenses

describeHomeRegionControlsResponse_nextToken :: Lens' DescribeHomeRegionControlsResponse (Maybe Text)

If a NextToken was returned by a previous call, more results are available. To retrieve the next page of results, make the call again using the returned token in NextToken.

describeHomeRegionControlsResponse_homeRegionControls :: Lens' DescribeHomeRegionControlsResponse (Maybe [HomeRegionControl])

An array that contains your HomeRegionControl objects.

describeHomeRegionControlsResponse_httpStatus :: Lens' DescribeHomeRegionControlsResponse Int

The response's http status code.