Amazonka.MigrationHubConfig.Types.HomeRegionControl

Description

Documentation

data HomeRegionControl

A home region control is an object that specifies the home region for an account, with some additional information. It contains a target (always of type ACCOUNT), an ID, and a time at which the home region was set.

See: newHomeRegionControl smart constructor.

Constructors

HomeRegionControl'

Fields

requestedTime :: Maybe POSIX

A timestamp representing the time when the customer called CreateHomeregionControl and set the home region for the account.

homeRegion :: Maybe Text

The AWS Region that's been set as home region. For example, "us-west-2" or "eu-central-1" are valid home regions.

controlId :: Maybe Text

A unique identifier that's generated for each home region control. It's always a string that begins with "hrc-" followed by 12 lowercase letters and numbers.

target :: Maybe Target

The target parameter specifies the identifier to which the home region is applied, which is always an ACCOUNT. It applies the home region to the current ACCOUNT.

newHomeRegionControl :: HomeRegionControl

Create a value of HomeRegionControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestedTime:HomeRegionControl', homeRegionControl_requestedTime - A timestamp representing the time when the customer called CreateHomeregionControl and set the home region for the account.

$sel:homeRegion:HomeRegionControl', homeRegionControl_homeRegion - The AWS Region that's been set as home region. For example, "us-west-2" or "eu-central-1" are valid home regions.

$sel:controlId:HomeRegionControl', homeRegionControl_controlId - A unique identifier that's generated for each home region control. It's always a string that begins with "hrc-" followed by 12 lowercase letters and numbers.

$sel:target:HomeRegionControl', homeRegionControl_target - The target parameter specifies the identifier to which the home region is applied, which is always an ACCOUNT. It applies the home region to the current ACCOUNT.

homeRegionControl_requestedTime :: Lens' HomeRegionControl (Maybe UTCTime)

A timestamp representing the time when the customer called CreateHomeregionControl and set the home region for the account.

homeRegionControl_homeRegion :: Lens' HomeRegionControl (Maybe Text)

The AWS Region that's been set as home region. For example, "us-west-2" or "eu-central-1" are valid home regions.

homeRegionControl_controlId :: Lens' HomeRegionControl (Maybe Text)

A unique identifier that's generated for each home region control. It's always a string that begins with "hrc-" followed by 12 lowercase letters and numbers.

homeRegionControl_target :: Lens' HomeRegionControl (Maybe Target)

The target parameter specifies the identifier to which the home region is applied, which is always an ACCOUNT. It applies the home region to the current ACCOUNT.