Amazonka.Mobile.CreateProject

Description

Creates an AWS Mobile Hub project.

Creating a Request

data CreateProject

Request structure used to request a project be created.

See: newCreateProject smart constructor.

Constructors

CreateProject'

Fields

contents :: Maybe ByteString

ZIP or YAML file which contains configuration settings to be used when creating the project. This may be the contents of the file downloaded from the URL provided in an export project operation.

name :: Maybe Text

Name of the project.

region :: Maybe Text

Default region where project resources should be created.

snapshotId :: Maybe Text

Unique identifier for an exported snapshot of project configuration. This snapshot identifier is included in the share URL when a project is exported.

newCreateProject :: CreateProject

Create a value of CreateProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contents:CreateProject', createProject_contents - ZIP or YAML file which contains configuration settings to be used when creating the project. This may be the contents of the file downloaded from the URL provided in an export project operation.

$sel:name:CreateProject', createProject_name - Name of the project.

$sel:region:CreateProject', createProject_region - Default region where project resources should be created.

$sel:snapshotId:CreateProject', createProject_snapshotId - Unique identifier for an exported snapshot of project configuration. This snapshot identifier is included in the share URL when a project is exported.

Request Lenses

createProject_contents :: Lens' CreateProject (Maybe ByteString)

ZIP or YAML file which contains configuration settings to be used when creating the project. This may be the contents of the file downloaded from the URL provided in an export project operation.

createProject_name :: Lens' CreateProject (Maybe Text)

Name of the project.

createProject_region :: Lens' CreateProject (Maybe Text)

Default region where project resources should be created.

createProject_snapshotId :: Lens' CreateProject (Maybe Text)

Unique identifier for an exported snapshot of project configuration. This snapshot identifier is included in the share URL when a project is exported.

Destructuring the Response

data CreateProjectResponse

Result structure used in response to a request to create a project.

See: newCreateProjectResponse smart constructor.

Constructors

CreateProjectResponse'

Fields

details :: Maybe ProjectDetails

Detailed information about the created AWS Mobile Hub project.

httpStatus :: Int

The response's http status code.

newCreateProjectResponse

Arguments

:: Int
$sel:httpStatus:CreateProjectResponse'

-> CreateProjectResponse

Create a value of CreateProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:details:CreateProjectResponse', createProjectResponse_details - Detailed information about the created AWS Mobile Hub project.

$sel:httpStatus:CreateProjectResponse', createProjectResponse_httpStatus - The response's http status code.

Response Lenses

createProjectResponse_details :: Lens' CreateProjectResponse (Maybe ProjectDetails)

Detailed information about the created AWS Mobile Hub project.

createProjectResponse_httpStatus :: Lens' CreateProjectResponse Int

The response's http status code.