Amazonka.Mobile.DescribeBundle

Description

Get the bundle details for the requested bundle id.

Creating a Request

data DescribeBundle

Request structure to request the details of a specific bundle.

See: newDescribeBundle smart constructor.

Constructors

DescribeBundle'

Fields

bundleId :: Text

Unique bundle identifier.

newDescribeBundle

Arguments

:: Text
$sel:bundleId:DescribeBundle'

-> DescribeBundle

Create a value of DescribeBundle with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bundleId:DescribeBundle', describeBundle_bundleId - Unique bundle identifier.

Request Lenses

describeBundle_bundleId :: Lens' DescribeBundle Text

Unique bundle identifier.

Destructuring the Response

data DescribeBundleResponse

Result structure contains the details of the bundle.

See: newDescribeBundleResponse smart constructor.

Constructors

DescribeBundleResponse'

Fields

details :: Maybe BundleDetails

The details of the bundle.

httpStatus :: Int

The response's http status code.

newDescribeBundleResponse

Arguments

:: Int
$sel:httpStatus:DescribeBundleResponse'

-> DescribeBundleResponse

Create a value of DescribeBundleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:details:DescribeBundleResponse', describeBundleResponse_details - The details of the bundle.

$sel:httpStatus:DescribeBundleResponse', describeBundleResponse_httpStatus - The response's http status code.

Response Lenses

describeBundleResponse_details :: Lens' DescribeBundleResponse (Maybe BundleDetails)

The details of the bundle.

describeBundleResponse_httpStatus :: Lens' DescribeBundleResponse Int

The response's http status code.