Amazonka.Mobile.Types

Description

Service Configuration

defaultService :: Service

API version 2017-07-01 of the Amazon Mobile SDK configuration.

Errors

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

No entity can be found with the specified identifier.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Too many requests have been received for this AWS account in too short a time. The request should be retried after some time delay.

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The service has encountered an unexpected error condition which prevents it from servicing the request.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable. The request should be retried after some time delay.

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

Credentials of the caller are insufficient to authorize the request.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request cannot be processed because some parameter is not valid or the project state prevents the operation from being performed.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

There are too many AWS Mobile Hub projects in the account or the account has exceeded the maximum number of resources in some AWS service. You should create another sub-account using AWS Organizations or remove some resources and retry your request.

_AccountActionRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

Account Action is required in order to continue the request.

Platform

newtype Platform

Developer desktop or target mobile app or website platform.

Constructors

Platform'

Fields

fromPlatform :: Text

Bundled Patterns

pattern Platform_ANDROID :: Platform

pattern Platform_JAVASCRIPT :: Platform

pattern Platform_LINUX :: Platform

pattern Platform_OBJC :: Platform

pattern Platform_OSX :: Platform

pattern Platform_SWIFT :: Platform

pattern Platform_WINDOWS :: Platform

ProjectState

newtype ProjectState

Synchronization state for a project.

Constructors

ProjectState'

Fields

fromProjectState :: Text

Bundled Patterns

pattern ProjectState_IMPORTING :: ProjectState

pattern ProjectState_NORMAL :: ProjectState

pattern ProjectState_SYNCING :: ProjectState

BundleDetails

data BundleDetails

The details of the bundle.

See: newBundleDetails smart constructor.

Constructors

BundleDetails'

Fields

availablePlatforms :: Maybe [Platform]

bundleId :: Maybe Text

version :: Maybe Text

iconUrl :: Maybe Text

title :: Maybe Text

description :: Maybe Text

newBundleDetails :: BundleDetails

Create a value of BundleDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availablePlatforms:BundleDetails', bundleDetails_availablePlatforms - Undocumented member.

$sel:bundleId:BundleDetails', bundleDetails_bundleId - Undocumented member.

$sel:version:BundleDetails', bundleDetails_version - Undocumented member.

$sel:iconUrl:BundleDetails', bundleDetails_iconUrl - Undocumented member.

$sel:title:BundleDetails', bundleDetails_title - Undocumented member.

$sel:description:BundleDetails', bundleDetails_description - Undocumented member.

bundleDetails_availablePlatforms :: Lens' BundleDetails (Maybe [Platform])

Undocumented member.

bundleDetails_bundleId :: Lens' BundleDetails (Maybe Text)

Undocumented member.

bundleDetails_version :: Lens' BundleDetails (Maybe Text)

Undocumented member.

bundleDetails_iconUrl :: Lens' BundleDetails (Maybe Text)

Undocumented member.

bundleDetails_title :: Lens' BundleDetails (Maybe Text)

Undocumented member.

bundleDetails_description :: Lens' BundleDetails (Maybe Text)

Undocumented member.

ProjectDetails

data ProjectDetails

Detailed information about an AWS Mobile Hub project.

See: newProjectDetails smart constructor.

Constructors

ProjectDetails'

Fields

state :: Maybe ProjectState

resources :: Maybe [Resource]

createdDate :: Maybe POSIX

Date the project was created.

consoleUrl :: Maybe Text

Website URL for this project in the AWS Mobile Hub console.

name :: Maybe Text

region :: Maybe Text

projectId :: Maybe Text

lastUpdatedDate :: Maybe POSIX

Date of the last modification of the project.

newProjectDetails :: ProjectDetails

Create a value of ProjectDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ProjectDetails', projectDetails_state - Undocumented member.

$sel:resources:ProjectDetails', projectDetails_resources - Undocumented member.

$sel:createdDate:ProjectDetails', projectDetails_createdDate - Date the project was created.

$sel:consoleUrl:ProjectDetails', projectDetails_consoleUrl - Website URL for this project in the AWS Mobile Hub console.

$sel:name:ProjectDetails', projectDetails_name - Undocumented member.

$sel:region:ProjectDetails', projectDetails_region - Undocumented member.

$sel:projectId:ProjectDetails', projectDetails_projectId - Undocumented member.

$sel:lastUpdatedDate:ProjectDetails', projectDetails_lastUpdatedDate - Date of the last modification of the project.

projectDetails_state :: Lens' ProjectDetails (Maybe ProjectState)

Undocumented member.

projectDetails_resources :: Lens' ProjectDetails (Maybe [Resource])

Undocumented member.

projectDetails_createdDate :: Lens' ProjectDetails (Maybe UTCTime)

Date the project was created.

projectDetails_consoleUrl :: Lens' ProjectDetails (Maybe Text)

Website URL for this project in the AWS Mobile Hub console.

projectDetails_name :: Lens' ProjectDetails (Maybe Text)

Undocumented member.

projectDetails_region :: Lens' ProjectDetails (Maybe Text)

Undocumented member.

projectDetails_projectId :: Lens' ProjectDetails (Maybe Text)

Undocumented member.

projectDetails_lastUpdatedDate :: Lens' ProjectDetails (Maybe UTCTime)

Date of the last modification of the project.

ProjectSummary

data ProjectSummary

Summary information about an AWS Mobile Hub project.

See: newProjectSummary smart constructor.

Constructors

ProjectSummary'

Fields

name :: Maybe Text

Name of the project.

projectId :: Maybe Text

Unique project identifier.

newProjectSummary :: ProjectSummary

Create a value of ProjectSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ProjectSummary', projectSummary_name - Name of the project.

$sel:projectId:ProjectSummary', projectSummary_projectId - Unique project identifier.

projectSummary_name :: Lens' ProjectSummary (Maybe Text)

Name of the project.

projectSummary_projectId :: Lens' ProjectSummary (Maybe Text)

Unique project identifier.

Resource

data Resource

Information about an instance of an AWS resource associated with a project.

See: newResource smart constructor.

Constructors

Resource'

Fields

feature :: Maybe Text

arn :: Maybe Text

name :: Maybe Text

attributes :: Maybe (HashMap Text Text)

type' :: Maybe Text

newResource :: Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:feature:Resource', resource_feature - Undocumented member.

$sel:arn:Resource', resource_arn - Undocumented member.

$sel:name:Resource', resource_name - Undocumented member.

$sel:attributes:Resource', resource_attributes - Undocumented member.

$sel:type':Resource', resource_type - Undocumented member.

resource_feature :: Lens' Resource (Maybe Text)

Undocumented member.

resource_arn :: Lens' Resource (Maybe Text)

Undocumented member.

resource_name :: Lens' Resource (Maybe Text)

Undocumented member.

resource_attributes :: Lens' Resource (Maybe (HashMap Text Text))

Undocumented member.

resource_type :: Lens' Resource (Maybe Text)

Undocumented member.