Amazonka.MwAA

Description

Derived from API version 2020-07-01 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Managed Workflows for Apache Airflow

This section contains the Amazon Managed Workflows for Apache Airflow (MWAA) API reference documentation. For more information, see What Is Amazon MWAA?.

Service Configuration

defaultService :: Service

API version 2020-07-01 of the Amazon MWAA SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MwAA.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

ValidationException: The provided input is not valid.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Access to the Airflow Web UI or CLI has been Denied. Please follow the MWAA user guide to setup permissions to access the Web UI and CLI functionality.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

InternalServerException: An internal error has occurred.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

ResourceNotFoundException: The resource is not available.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListEnvironments (Paginated)

data ListEnvironments

See: newListEnvironments smart constructor.

Constructors

ListEnvironments' (Maybe Text) (Maybe Natural)

newListEnvironments :: ListEnvironments

Create a value of ListEnvironments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEnvironments', listEnvironments_nextToken - Retrieves the next page of the results.

$sel:maxResults:ListEnvironments', listEnvironments_maxResults - The maximum number of results to retrieve per page. For example, 5 environments per page.

data ListEnvironmentsResponse

See: newListEnvironmentsResponse smart constructor.

Constructors

ListEnvironmentsResponse' (Maybe Text) Int [Text]

newListEnvironmentsResponse

Arguments

:: Int
$sel:httpStatus:ListEnvironmentsResponse'

-> ListEnvironmentsResponse

Create a value of ListEnvironmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEnvironments', listEnvironmentsResponse_nextToken - Retrieves the next page of the results.

$sel:httpStatus:ListEnvironmentsResponse', listEnvironmentsResponse_httpStatus - The response's http status code.

$sel:environments:ListEnvironmentsResponse', listEnvironmentsResponse_environments - Returns the list of Amazon MWAA environments.

UpdateEnvironment

data UpdateEnvironment

See: newUpdateEnvironment smart constructor.

Constructors

UpdateEnvironment' (Maybe Int) (Maybe Natural) (Maybe Text) (Maybe WebserverAccessMode) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (Sensitive (HashMap Text Text))) (Maybe LoggingConfigurationInput) (Maybe Text) (Maybe UpdateNetworkConfigurationInput) (Maybe Text) (Maybe Natural) Text

newUpdateEnvironment

Arguments

:: Text
$sel:name:UpdateEnvironment'

-> UpdateEnvironment

Create a value of UpdateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schedulers:UpdateEnvironment', updateEnvironment_schedulers - The number of Apache Airflow schedulers to run in your Amazon MWAA environment.

$sel:minWorkers:UpdateEnvironment', updateEnvironment_minWorkers - The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

$sel:pluginsS3Path:UpdateEnvironment', updateEnvironment_pluginsS3Path - The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

$sel:webserverAccessMode:UpdateEnvironment', updateEnvironment_webserverAccessMode - The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

$sel:airflowVersion:UpdateEnvironment', updateEnvironment_airflowVersion - The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

$sel:weeklyMaintenanceWindowStart:UpdateEnvironment', updateEnvironment_weeklyMaintenanceWindowStart - The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following:

MON|TUE|WED|THU|FRI|SAT|SUN:([01]\\d|2[0-3]):(00|30)

$sel:executionRoleArn:UpdateEnvironment', updateEnvironment_executionRoleArn - The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

$sel:requirementsS3ObjectVersion:UpdateEnvironment', updateEnvironment_requirementsS3ObjectVersion - The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

$sel:sourceBucketArn:UpdateEnvironment', updateEnvironment_sourceBucketArn - The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

$sel:dagS3Path:UpdateEnvironment', updateEnvironment_dagS3Path - The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

$sel:pluginsS3ObjectVersion:UpdateEnvironment', updateEnvironment_pluginsS3ObjectVersion - The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

$sel:airflowConfigurationOptions:UpdateEnvironment', updateEnvironment_airflowConfigurationOptions - A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

$sel:loggingConfiguration:UpdateEnvironment', updateEnvironment_loggingConfiguration - Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

$sel:environmentClass:UpdateEnvironment', updateEnvironment_environmentClass - The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

$sel:networkConfiguration:UpdateEnvironment', updateEnvironment_networkConfiguration - The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

$sel:requirementsS3Path:UpdateEnvironment', updateEnvironment_requirementsS3Path - The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

$sel:maxWorkers:UpdateEnvironment', updateEnvironment_maxWorkers - The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

$sel:name:UpdateEnvironment', updateEnvironment_name - The name of your Amazon MWAA environment. For example, MyMWAAEnvironment.

data UpdateEnvironmentResponse

See: newUpdateEnvironmentResponse smart constructor.

Constructors

UpdateEnvironmentResponse' (Maybe Text) Int

newUpdateEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:UpdateEnvironmentResponse'

-> UpdateEnvironmentResponse

Create a value of UpdateEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:UpdateEnvironmentResponse', updateEnvironmentResponse_arn - The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

$sel:httpStatus:UpdateEnvironmentResponse', updateEnvironmentResponse_httpStatus - The response's http status code.

DeleteEnvironment

data DeleteEnvironment

See: newDeleteEnvironment smart constructor.

Constructors

DeleteEnvironment' Text

newDeleteEnvironment

Arguments

:: Text
$sel:name:DeleteEnvironment'

-> DeleteEnvironment

Create a value of DeleteEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteEnvironment', deleteEnvironment_name - The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

data DeleteEnvironmentResponse

See: newDeleteEnvironmentResponse smart constructor.

Constructors

DeleteEnvironmentResponse' Int

newDeleteEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:DeleteEnvironmentResponse'

-> DeleteEnvironmentResponse

Create a value of DeleteEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteEnvironmentResponse', deleteEnvironmentResponse_httpStatus - The response's http status code.

CreateWebLoginToken

data CreateWebLoginToken

See: newCreateWebLoginToken smart constructor.

Constructors

CreateWebLoginToken' Text

newCreateWebLoginToken

Arguments

:: Text
$sel:name:CreateWebLoginToken'

-> CreateWebLoginToken

Create a value of CreateWebLoginToken with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateWebLoginToken', createWebLoginToken_name - Create an Airflow Web UI login token request for a MWAA environment.

data CreateWebLoginTokenResponse

See: newCreateWebLoginTokenResponse smart constructor.

Constructors

CreateWebLoginTokenResponse' (Maybe Text) (Maybe (Sensitive Text)) Int

newCreateWebLoginTokenResponse

Arguments

:: Int
$sel:httpStatus:CreateWebLoginTokenResponse'

-> CreateWebLoginTokenResponse

Create a value of CreateWebLoginTokenResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:webServerHostname:CreateWebLoginTokenResponse', createWebLoginTokenResponse_webServerHostname - Create an Airflow Web UI login token response for the provided webserver hostname.

$sel:webToken:CreateWebLoginTokenResponse', createWebLoginTokenResponse_webToken - Create an Airflow Web UI login token response for the provided JWT token.

$sel:httpStatus:CreateWebLoginTokenResponse', createWebLoginTokenResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The key-value tag pairs associated to your environment. To learn more, see Tagging AWS resources.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

$sel:tags:TagResource', tagResource_tags - The key-value tag pairs you want to associate to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

GetEnvironment

data GetEnvironment

See: newGetEnvironment smart constructor.

Constructors

GetEnvironment' Text

newGetEnvironment

Arguments

:: Text
$sel:name:GetEnvironment'

-> GetEnvironment

Create a value of GetEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetEnvironment', getEnvironment_name - The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

data GetEnvironmentResponse

See: newGetEnvironmentResponse smart constructor.

Constructors

GetEnvironmentResponse' (Maybe Environment) Int

newGetEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:GetEnvironmentResponse'

-> GetEnvironmentResponse

Create a value of GetEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environment:GetEnvironmentResponse', getEnvironmentResponse_environment - An object containing all available details about the environment.

$sel:httpStatus:GetEnvironmentResponse', getEnvironmentResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The key-value tag pair you want to remove. For example, "Environment": "Staging".

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

PublishMetrics

data PublishMetrics

See: newPublishMetrics smart constructor.

Constructors

PublishMetrics' Text [MetricDatum]

newPublishMetrics

Arguments

:: Text
$sel:environmentName:PublishMetrics'

-> PublishMetrics

Create a value of PublishMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentName:PublishMetrics', publishMetrics_environmentName - Publishes environment metric data to Amazon CloudWatch.

$sel:metricData:PublishMetrics', publishMetrics_metricData - Publishes metric data points to Amazon CloudWatch. CloudWatch associates the data points with the specified metrica.

data PublishMetricsResponse

See: newPublishMetricsResponse smart constructor.

Constructors

PublishMetricsResponse' Int

newPublishMetricsResponse

Arguments

:: Int
$sel:httpStatus:PublishMetricsResponse'

-> PublishMetricsResponse

Create a value of PublishMetricsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PublishMetricsResponse', publishMetricsResponse_httpStatus - The response's http status code.

CreateEnvironment

data CreateEnvironment

This section contains the Amazon Managed Workflows for Apache Airflow (MWAA) API reference documentation to create an environment. For more information, see Get started with Amazon Managed Workflows for Apache Airflow.

See: newCreateEnvironment smart constructor.

Constructors

CreateEnvironment' (Maybe Int) (Maybe Natural) (Maybe Text) (Maybe WebserverAccessMode) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (Sensitive (HashMap Text Text))) (Maybe LoggingConfigurationInput) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe Text) (Maybe Natural) Text Text Text NetworkConfiguration Text

newCreateEnvironment

Arguments

:: Text
$sel:dagS3Path:CreateEnvironment'

-> Text
$sel:executionRoleArn:CreateEnvironment'

-> Text
$sel:name:CreateEnvironment'

-> NetworkConfiguration
$sel:networkConfiguration:CreateEnvironment'

-> Text
$sel:sourceBucketArn:CreateEnvironment'

-> CreateEnvironment

Create a value of CreateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schedulers:CreateEnvironment', createEnvironment_schedulers - The number of Apache Airflow schedulers to run in your environment.

$sel:minWorkers:CreateEnvironment', createEnvironment_minWorkers - The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

$sel:pluginsS3Path:CreateEnvironment', createEnvironment_pluginsS3Path - The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

$sel:webserverAccessMode:CreateEnvironment', createEnvironment_webserverAccessMode - The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

$sel:airflowVersion:CreateEnvironment', createEnvironment_airflowVersion - The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

$sel:kmsKey:CreateEnvironment', createEnvironment_kmsKey - The AWS Key Management Service (KMS) key to encrypt the data in your environment. You can use an AWS owned CMK, or a Customer managed CMK (advanced). To learn more, see Get started with Amazon Managed Workflows for Apache Airflow.

$sel:weeklyMaintenanceWindowStart:CreateEnvironment', createEnvironment_weeklyMaintenanceWindowStart - The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following:

MON|TUE|WED|THU|FRI|SAT|SUN:([01]\\d|2[0-3]):(00|30)

$sel:requirementsS3ObjectVersion:CreateEnvironment', createEnvironment_requirementsS3ObjectVersion - The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

$sel:pluginsS3ObjectVersion:CreateEnvironment', createEnvironment_pluginsS3ObjectVersion - The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

$sel:airflowConfigurationOptions:CreateEnvironment', createEnvironment_airflowConfigurationOptions - A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

$sel:loggingConfiguration:CreateEnvironment', createEnvironment_loggingConfiguration - Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

$sel:environmentClass:CreateEnvironment', createEnvironment_environmentClass - The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

$sel:tags:CreateEnvironment', createEnvironment_tags - The key-value tag pairs you want to associate to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

$sel:requirementsS3Path:CreateEnvironment', createEnvironment_requirementsS3Path - The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

$sel:maxWorkers:CreateEnvironment', createEnvironment_maxWorkers - The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

$sel:dagS3Path:CreateEnvironment', createEnvironment_dagS3Path - The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

$sel:executionRoleArn:CreateEnvironment', createEnvironment_executionRoleArn - The Amazon Resource Name (ARN) of the execution role for your environment. An execution role is an AWS Identity and Access Management (IAM) role that grants MWAA permission to access AWS services and resources used by your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

$sel:name:CreateEnvironment', createEnvironment_name - The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

$sel:networkConfiguration:CreateEnvironment', createEnvironment_networkConfiguration - The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

$sel:sourceBucketArn:CreateEnvironment', createEnvironment_sourceBucketArn - The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

data CreateEnvironmentResponse

See: newCreateEnvironmentResponse smart constructor.

Constructors

CreateEnvironmentResponse' (Maybe Text) Int

newCreateEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:CreateEnvironmentResponse'

-> CreateEnvironmentResponse

Create a value of CreateEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateEnvironmentResponse', createEnvironmentResponse_arn - The Amazon Resource Name (ARN) returned in the response for the environment.

$sel:httpStatus:CreateEnvironmentResponse', createEnvironmentResponse_httpStatus - The response's http status code.

CreateCliToken

data CreateCliToken

See: newCreateCliToken smart constructor.

Constructors

CreateCliToken' Text

newCreateCliToken

Arguments

:: Text
$sel:name:CreateCliToken'

-> CreateCliToken

Create a value of CreateCliToken with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateCliToken', createCliToken_name - Create a CLI token request for a MWAA environment.

data CreateCliTokenResponse

See: newCreateCliTokenResponse smart constructor.

Constructors

CreateCliTokenResponse' (Maybe Text) (Maybe (Sensitive Text)) Int

newCreateCliTokenResponse

Arguments

:: Int
$sel:httpStatus:CreateCliTokenResponse'

-> CreateCliTokenResponse

Create a value of CreateCliTokenResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:webServerHostname:CreateCliTokenResponse', createCliTokenResponse_webServerHostname - Create an Airflow CLI login token response for the provided webserver hostname.

$sel:cliToken:CreateCliTokenResponse', createCliTokenResponse_cliToken - Create an Airflow CLI login token response for the provided JWT token.

$sel:httpStatus:CreateCliTokenResponse', createCliTokenResponse_httpStatus - The response's http status code.

Types

EnvironmentStatus

newtype EnvironmentStatus

Constructors

EnvironmentStatus'

Fields

fromEnvironmentStatus :: Text

Bundled Patterns

pattern EnvironmentStatus_AVAILABLE :: EnvironmentStatus

pattern EnvironmentStatus_CREATE_FAILED :: EnvironmentStatus

pattern EnvironmentStatus_CREATING :: EnvironmentStatus

pattern EnvironmentStatus_DELETED :: EnvironmentStatus

pattern EnvironmentStatus_DELETING :: EnvironmentStatus

pattern EnvironmentStatus_UNAVAILABLE :: EnvironmentStatus

pattern EnvironmentStatus_UPDATE_FAILED :: EnvironmentStatus

pattern EnvironmentStatus_UPDATING :: EnvironmentStatus

LoggingLevel

newtype LoggingLevel

Constructors

LoggingLevel'

Fields

fromLoggingLevel :: Text

Bundled Patterns

pattern LoggingLevel_CRITICAL :: LoggingLevel

pattern LoggingLevel_DEBUG :: LoggingLevel

pattern LoggingLevel_ERROR :: LoggingLevel

pattern LoggingLevel_INFO :: LoggingLevel

pattern LoggingLevel_WARNING :: LoggingLevel

Unit

newtype Unit

Unit

Constructors

Unit'

Fields

fromUnit :: Text

Bundled Patterns

pattern Unit_Bits :: Unit

pattern Unit_Bits_Second :: Unit

pattern Unit_Bytes :: Unit

pattern Unit_Bytes_Second :: Unit

pattern Unit_Count :: Unit

pattern Unit_Count_Second :: Unit

pattern Unit_Gigabits :: Unit

pattern Unit_Gigabits_Second :: Unit

pattern Unit_Gigabytes :: Unit

pattern Unit_Gigabytes_Second :: Unit

pattern Unit_Kilobits :: Unit

pattern Unit_Kilobits_Second :: Unit

pattern Unit_Kilobytes :: Unit

pattern Unit_Kilobytes_Second :: Unit

pattern Unit_Megabits :: Unit

pattern Unit_Megabits_Second :: Unit

pattern Unit_Megabytes :: Unit

pattern Unit_Megabytes_Second :: Unit

pattern Unit_Microseconds :: Unit

pattern Unit_Milliseconds :: Unit

pattern Unit_None :: Unit

pattern Unit_Percent :: Unit

pattern Unit_Seconds :: Unit

pattern Unit_Terabits :: Unit

pattern Unit_Terabits_Second :: Unit

pattern Unit_Terabytes :: Unit

pattern Unit_Terabytes_Second :: Unit

UpdateStatus

newtype UpdateStatus

Constructors

UpdateStatus'

Fields

fromUpdateStatus :: Text

Bundled Patterns

pattern UpdateStatus_FAILED :: UpdateStatus

pattern UpdateStatus_PENDING :: UpdateStatus

pattern UpdateStatus_SUCCESS :: UpdateStatus

WebserverAccessMode

newtype WebserverAccessMode

Constructors

WebserverAccessMode'

Fields

fromWebserverAccessMode :: Text

Bundled Patterns

pattern WebserverAccessMode_PRIVATE_ONLY :: WebserverAccessMode

pattern WebserverAccessMode_PUBLIC_ONLY :: WebserverAccessMode

Dimension

data Dimension

Internal only API.

See: newDimension smart constructor.

Constructors

Dimension' Text Text

newDimension

Arguments

:: Text
$sel:name:Dimension'

-> Text
$sel:value:Dimension'

-> Dimension

Create a value of Dimension with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Dimension', dimension_name - Internal only API.

$sel:value:Dimension', dimension_value - Internal only API.

Environment

data Environment

The Amazon Managed Workflows for Apache Airflow (MWAA) environment.

See: newEnvironment smart constructor.

Constructors

Environment' (Maybe Text) (Maybe Int) (Maybe EnvironmentStatus) (Maybe Natural) (Maybe Text) (Maybe WebserverAccessMode) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe LastUpdate) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe LoggingConfiguration) (Maybe Text) (Maybe NetworkConfiguration) (Maybe (HashMap Text Text)) (Maybe Text) (Maybe Natural)

newEnvironment :: Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceRoleArn:Environment', environment_serviceRoleArn - The Amazon Resource Name (ARN) for the service-linked role of the environment. To learn more, see Amazon MWAA Service-linked role.

$sel:schedulers:Environment', environment_schedulers - The number of Apache Airflow schedulers that run in your Amazon MWAA environment.

$sel:status:Environment', environment_status - The status of the Amazon MWAA environment. Valid values:

CREATING - Indicates the request to create the environment is in progress.

CREATE_FAILED - Indicates the request to create the environment failed, and the environment could not be created.

AVAILABLE - Indicates the request was successful and the environment is ready to use.

UPDATING - Indicates the request to update the environment is in progress.

DELETING - Indicates the request to delete the environment is in progress.

DELETED - Indicates the request to delete the environment is complete, and the environment has been deleted.

UNAVAILABLE - Indicates the request failed, but the environment was unable to rollback and is not in a stable state.

UPDATE_FAILED - Indicates the request to update the environment failed, and the environment has rolled back successfully and is ready to use.

We recommend reviewing our troubleshooting guide for a list of common errors and their solutions. To learn more, see Amazon MWAA troubleshooting.

$sel:minWorkers:Environment', environment_minWorkers - The minimum number of workers that run in your environment. For example, 2.

$sel:pluginsS3Path:Environment', environment_pluginsS3Path - The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. To learn more, see Installing custom plugins.

$sel:webserverAccessMode:Environment', environment_webserverAccessMode - The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

$sel:airflowVersion:Environment', environment_airflowVersion - The Apache Airflow version on your environment. For example, v1.10.12.

$sel:kmsKey:Environment', environment_kmsKey - The Key Management Service (KMS) encryption key used to encrypt the data in your environment.

$sel:arn:Environment', environment_arn - The Amazon Resource Name (ARN) of the Amazon MWAA environment.

$sel:createdAt:Environment', environment_createdAt - The day and time the environment was created.

$sel:weeklyMaintenanceWindowStart:Environment', environment_weeklyMaintenanceWindowStart - The day and time of the week that weekly maintenance updates are scheduled. For example: TUE:03:30.

$sel:executionRoleArn:Environment', environment_executionRoleArn - The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

$sel:requirementsS3ObjectVersion:Environment', environment_requirementsS3ObjectVersion - The version of the requirements.txt file on your Amazon S3 bucket. To learn more, see Installing Python dependencies.

$sel:lastUpdate:Environment', environment_lastUpdate - Undocumented member.

$sel:sourceBucketArn:Environment', environment_sourceBucketArn - The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

$sel:webserverUrl:Environment', environment_webserverUrl - The Apache Airflow Web server host name for the Amazon MWAA environment. To learn more, see Accessing the Apache Airflow UI.

$sel:dagS3Path:Environment', environment_dagS3Path - The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

$sel:name:Environment', environment_name - The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

$sel:pluginsS3ObjectVersion:Environment', environment_pluginsS3ObjectVersion - The version of the plugins.zip file on your Amazon S3 bucket. To learn more, see Installing custom plugins.

$sel:airflowConfigurationOptions:Environment', environment_airflowConfigurationOptions - A list of key-value pairs containing the Apache Airflow configuration options attached to your environment. To learn more, see Apache Airflow configuration options.

$sel:loggingConfiguration:Environment', environment_loggingConfiguration - The Apache Airflow logs being sent to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

$sel:environmentClass:Environment', environment_environmentClass - The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

$sel:networkConfiguration:Environment', environment_networkConfiguration - Undocumented member.

$sel:tags:Environment', environment_tags - The key-value tag pairs associated to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

$sel:requirementsS3Path:Environment', environment_requirementsS3Path - The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. To learn more, see Installing Python dependencies.

$sel:maxWorkers:Environment', environment_maxWorkers - The maximum number of workers that run in your environment. For example, 20.

LastUpdate

data LastUpdate

The status of the last update on the environment, and any errors that were encountered.

See: newLastUpdate smart constructor.

Constructors

LastUpdate' (Maybe UpdateStatus) (Maybe POSIX) (Maybe UpdateError)

newLastUpdate :: LastUpdate

Create a value of LastUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:LastUpdate', lastUpdate_status - The status of the last update on the environment. Valid values: SUCCESS, PENDING, FAILED.

$sel:createdAt:LastUpdate', lastUpdate_createdAt - The day and time of the last update on the environment.

$sel:error:LastUpdate', lastUpdate_error - The error that was encountered during the last update of the environment.

LoggingConfiguration

data LoggingConfiguration

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

See: newLoggingConfiguration smart constructor.

Constructors

LoggingConfiguration' (Maybe ModuleLoggingConfiguration) (Maybe ModuleLoggingConfiguration) (Maybe ModuleLoggingConfiguration) (Maybe ModuleLoggingConfiguration) (Maybe ModuleLoggingConfiguration)

newLoggingConfiguration :: LoggingConfiguration

Create a value of LoggingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskLogs:LoggingConfiguration', loggingConfiguration_taskLogs - Undocumented member.

$sel:webserverLogs:LoggingConfiguration', loggingConfiguration_webserverLogs - Undocumented member.

$sel:schedulerLogs:LoggingConfiguration', loggingConfiguration_schedulerLogs - Undocumented member.

$sel:dagProcessingLogs:LoggingConfiguration', loggingConfiguration_dagProcessingLogs - Undocumented member.

$sel:workerLogs:LoggingConfiguration', loggingConfiguration_workerLogs - Undocumented member.

LoggingConfigurationInput

data LoggingConfigurationInput

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

See: newLoggingConfigurationInput smart constructor.

Constructors

LoggingConfigurationInput' (Maybe ModuleLoggingConfigurationInput) (Maybe ModuleLoggingConfigurationInput) (Maybe ModuleLoggingConfigurationInput) (Maybe ModuleLoggingConfigurationInput) (Maybe ModuleLoggingConfigurationInput)

newLoggingConfigurationInput :: LoggingConfigurationInput

Create a value of LoggingConfigurationInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskLogs:LoggingConfigurationInput', loggingConfigurationInput_taskLogs - Undocumented member.

$sel:webserverLogs:LoggingConfigurationInput', loggingConfigurationInput_webserverLogs - Undocumented member.

$sel:schedulerLogs:LoggingConfigurationInput', loggingConfigurationInput_schedulerLogs - Undocumented member.

$sel:dagProcessingLogs:LoggingConfigurationInput', loggingConfigurationInput_dagProcessingLogs - Undocumented member.

$sel:workerLogs:LoggingConfigurationInput', loggingConfigurationInput_workerLogs - Undocumented member.

MetricDatum

data MetricDatum

Internal only API.

See: newMetricDatum smart constructor.

Constructors

MetricDatum' (Maybe Double) (Maybe [Dimension]) (Maybe Unit) (Maybe StatisticSet) Text POSIX

newMetricDatum

Arguments

:: Text
$sel:metricName:MetricDatum'

-> UTCTime
$sel:timestamp:MetricDatum'

-> MetricDatum

Create a value of MetricDatum with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:MetricDatum', metricDatum_value - Internal only API.

$sel:dimensions:MetricDatum', metricDatum_dimensions - Internal only API.

$sel:unit:MetricDatum', metricDatum_unit - Undocumented member.

$sel:statisticValues:MetricDatum', metricDatum_statisticValues - Internal only API.

$sel:metricName:MetricDatum', metricDatum_metricName - Internal only API.

$sel:timestamp:MetricDatum', metricDatum_timestamp - Internal only API.

ModuleLoggingConfiguration

data ModuleLoggingConfiguration

Defines the type of logs to send for the Apache Airflow log type (e.g. DagProcessingLogs). Valid values: CloudWatchLogGroupArn, Enabled, LogLevel.

See: newModuleLoggingConfiguration smart constructor.

Constructors

ModuleLoggingConfiguration' (Maybe LoggingLevel) (Maybe Bool) (Maybe Text)

newModuleLoggingConfiguration :: ModuleLoggingConfiguration

Create a value of ModuleLoggingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logLevel:ModuleLoggingConfiguration', moduleLoggingConfiguration_logLevel - Defines the Apache Airflow logs to send for the log type (e.g. DagProcessingLogs) to CloudWatch Logs. Valid values: CRITICAL, ERROR, WARNING, INFO.

$sel:enabled:ModuleLoggingConfiguration', moduleLoggingConfiguration_enabled - Indicates whether to enable the Apache Airflow log type (e.g. DagProcessingLogs) in CloudWatch Logs.

$sel:cloudWatchLogGroupArn:ModuleLoggingConfiguration', moduleLoggingConfiguration_cloudWatchLogGroupArn - The Amazon Resource Name (ARN) for the CloudWatch Logs group where the Apache Airflow log type (e.g. DagProcessingLogs) is published. For example, arn:aws:logs:us-east-1:123456789012:log-group:airflow-MyMWAAEnvironment-MwaaEnvironment-DAGProcessing:*.

ModuleLoggingConfigurationInput

data ModuleLoggingConfigurationInput

Defines the type of logs to send for the Apache Airflow log type (e.g. DagProcessingLogs). Valid values: CloudWatchLogGroupArn, Enabled, LogLevel.

See: newModuleLoggingConfigurationInput smart constructor.

Constructors

ModuleLoggingConfigurationInput' Bool LoggingLevel

newModuleLoggingConfigurationInput

Arguments

:: Bool
$sel:enabled:ModuleLoggingConfigurationInput'

-> LoggingLevel
$sel:logLevel:ModuleLoggingConfigurationInput'

-> ModuleLoggingConfigurationInput

Create a value of ModuleLoggingConfigurationInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:ModuleLoggingConfigurationInput', moduleLoggingConfigurationInput_enabled - Indicates whether to enable the Apache Airflow log type (e.g. DagProcessingLogs) in CloudWatch Logs.

$sel:logLevel:ModuleLoggingConfigurationInput', moduleLoggingConfigurationInput_logLevel - Defines the Apache Airflow logs to send for the log type (e.g. DagProcessingLogs) to CloudWatch Logs. Valid values: CRITICAL, ERROR, WARNING, INFO.

NetworkConfiguration

data NetworkConfiguration

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

See: newNetworkConfiguration smart constructor.

Constructors

NetworkConfiguration' (Maybe (NonEmpty Text)) (Maybe (NonEmpty Text))

newNetworkConfiguration :: NetworkConfiguration

Create a value of NetworkConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:NetworkConfiguration', networkConfiguration_securityGroupIds - A list of 1 or more security group IDs. Accepts up to 5 security group IDs. A security group must be attached to the same VPC as the subnets. To learn more, see Security in your VPC on Amazon MWAA.

$sel:subnetIds:NetworkConfiguration', networkConfiguration_subnetIds - A list of 2 subnet IDs. Required to create an environment. Must be private subnets in two different availability zones. A subnet must be attached to the same VPC as the security group.

StatisticSet

data StatisticSet

Internal only API.

See: newStatisticSet smart constructor.

Constructors

StatisticSet' (Maybe Int) (Maybe Double) (Maybe Double) (Maybe Double)

newStatisticSet :: StatisticSet

Create a value of StatisticSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sampleCount:StatisticSet', statisticSet_sampleCount - Internal only API.

$sel:maximum:StatisticSet', statisticSet_maximum - Internal only API.

$sel:minimum:StatisticSet', statisticSet_minimum - Internal only API.

$sel:sum:StatisticSet', statisticSet_sum - Internal only API.

UpdateError

data UpdateError

An object containing the error encountered with the last update: ErrorCode, ErrorMessage.

See: newUpdateError smart constructor.

Constructors

UpdateError' (Maybe Text) (Maybe Text)

newUpdateError :: UpdateError

Create a value of UpdateError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:UpdateError', updateError_errorCode - The error code that corresponds to the error with the last update.

$sel:errorMessage:UpdateError', updateError_errorMessage - The error message that corresponds to the error code.

UpdateNetworkConfigurationInput

data UpdateNetworkConfigurationInput

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

See: newUpdateNetworkConfigurationInput smart constructor.

Constructors

UpdateNetworkConfigurationInput' (NonEmpty Text)

newUpdateNetworkConfigurationInput

Arguments

:: NonEmpty Text
$sel:securityGroupIds:UpdateNetworkConfigurationInput'

-> UpdateNetworkConfigurationInput

Create a value of UpdateNetworkConfigurationInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:UpdateNetworkConfigurationInput', updateNetworkConfigurationInput_securityGroupIds - A list of 1 or more security group IDs. Accepts up to 5 security group IDs. A security group must be attached to the same VPC as the subnets. To learn more, see Security in your VPC on Amazon MWAA.