Amazonka.MwAA.CreateEnvironment

Description

Creates an Amazon Managed Workflows for Apache Airflow (MWAA) environment.

Creating a Request

data CreateEnvironment

This section contains the Amazon Managed Workflows for Apache Airflow (MWAA) API reference documentation to create an environment. For more information, see Get started with Amazon Managed Workflows for Apache Airflow.

See: newCreateEnvironment smart constructor.

Constructors

CreateEnvironment'

Fields

schedulers :: Maybe Int

The number of Apache Airflow schedulers to run in your environment.

minWorkers :: Maybe Natural

The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

pluginsS3Path :: Maybe Text

The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

webserverAccessMode :: Maybe WebserverAccessMode

The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

airflowVersion :: Maybe Text

The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

kmsKey :: Maybe Text

The AWS Key Management Service (KMS) key to encrypt the data in your environment. You can use an AWS owned CMK, or a Customer managed CMK (advanced). To learn more, see Get started with Amazon Managed Workflows for Apache Airflow.

weeklyMaintenanceWindowStart :: Maybe Text

The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following:

MON|TUE|WED|THU|FRI|SAT|SUN:([01]\\d|2[0-3]):(00|30)

requirementsS3ObjectVersion :: Maybe Text

The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

pluginsS3ObjectVersion :: Maybe Text

The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

airflowConfigurationOptions :: Maybe (Sensitive (HashMap Text Text))

A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

loggingConfiguration :: Maybe LoggingConfigurationInput

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

environmentClass :: Maybe Text

The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

tags :: Maybe (HashMap Text Text)

The key-value tag pairs you want to associate to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

requirementsS3Path :: Maybe Text

The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

maxWorkers :: Maybe Natural

The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

dagS3Path :: Text

The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

executionRoleArn :: Text

The Amazon Resource Name (ARN) of the execution role for your environment. An execution role is an AWS Identity and Access Management (IAM) role that grants MWAA permission to access AWS services and resources used by your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

name :: Text

The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

networkConfiguration :: NetworkConfiguration

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

sourceBucketArn :: Text

The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

newCreateEnvironment

Arguments

:: Text
$sel:dagS3Path:CreateEnvironment'

-> Text
$sel:executionRoleArn:CreateEnvironment'

-> Text
$sel:name:CreateEnvironment'

-> NetworkConfiguration
$sel:networkConfiguration:CreateEnvironment'

-> Text
$sel:sourceBucketArn:CreateEnvironment'

-> CreateEnvironment

Create a value of CreateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schedulers:CreateEnvironment', createEnvironment_schedulers - The number of Apache Airflow schedulers to run in your environment.

$sel:minWorkers:CreateEnvironment', createEnvironment_minWorkers - The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

$sel:pluginsS3Path:CreateEnvironment', createEnvironment_pluginsS3Path - The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

$sel:webserverAccessMode:CreateEnvironment', createEnvironment_webserverAccessMode - The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

$sel:airflowVersion:CreateEnvironment', createEnvironment_airflowVersion - The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

$sel:kmsKey:CreateEnvironment', createEnvironment_kmsKey - The AWS Key Management Service (KMS) key to encrypt the data in your environment. You can use an AWS owned CMK, or a Customer managed CMK (advanced). To learn more, see Get started with Amazon Managed Workflows for Apache Airflow.

$sel:weeklyMaintenanceWindowStart:CreateEnvironment', createEnvironment_weeklyMaintenanceWindowStart - The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following:

MON|TUE|WED|THU|FRI|SAT|SUN:([01]\\d|2[0-3]):(00|30)

$sel:requirementsS3ObjectVersion:CreateEnvironment', createEnvironment_requirementsS3ObjectVersion - The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

$sel:pluginsS3ObjectVersion:CreateEnvironment', createEnvironment_pluginsS3ObjectVersion - The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

$sel:airflowConfigurationOptions:CreateEnvironment', createEnvironment_airflowConfigurationOptions - A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

$sel:loggingConfiguration:CreateEnvironment', createEnvironment_loggingConfiguration - Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

$sel:environmentClass:CreateEnvironment', createEnvironment_environmentClass - The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

$sel:tags:CreateEnvironment', createEnvironment_tags - The key-value tag pairs you want to associate to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

$sel:requirementsS3Path:CreateEnvironment', createEnvironment_requirementsS3Path - The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

$sel:maxWorkers:CreateEnvironment', createEnvironment_maxWorkers - The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

$sel:dagS3Path:CreateEnvironment', createEnvironment_dagS3Path - The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

$sel:executionRoleArn:CreateEnvironment', createEnvironment_executionRoleArn - The Amazon Resource Name (ARN) of the execution role for your environment. An execution role is an AWS Identity and Access Management (IAM) role that grants MWAA permission to access AWS services and resources used by your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

$sel:name:CreateEnvironment', createEnvironment_name - The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

$sel:networkConfiguration:CreateEnvironment', createEnvironment_networkConfiguration - The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

$sel:sourceBucketArn:CreateEnvironment', createEnvironment_sourceBucketArn - The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

Request Lenses

createEnvironment_schedulers :: Lens' CreateEnvironment (Maybe Int)

The number of Apache Airflow schedulers to run in your environment.

createEnvironment_minWorkers :: Lens' CreateEnvironment (Maybe Natural)

The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

createEnvironment_pluginsS3Path :: Lens' CreateEnvironment (Maybe Text)

The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

createEnvironment_webserverAccessMode :: Lens' CreateEnvironment (Maybe WebserverAccessMode)

The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

createEnvironment_airflowVersion :: Lens' CreateEnvironment (Maybe Text)

The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

createEnvironment_kmsKey :: Lens' CreateEnvironment (Maybe Text)

The AWS Key Management Service (KMS) key to encrypt the data in your environment. You can use an AWS owned CMK, or a Customer managed CMK (advanced). To learn more, see Get started with Amazon Managed Workflows for Apache Airflow.

createEnvironment_weeklyMaintenanceWindowStart :: Lens' CreateEnvironment (Maybe Text)

The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following:

MON|TUE|WED|THU|FRI|SAT|SUN:([01]\\d|2[0-3]):(00|30)

createEnvironment_requirementsS3ObjectVersion :: Lens' CreateEnvironment (Maybe Text)

The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

createEnvironment_pluginsS3ObjectVersion :: Lens' CreateEnvironment (Maybe Text)

The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

createEnvironment_airflowConfigurationOptions :: Lens' CreateEnvironment (Maybe (HashMap Text Text))

A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

createEnvironment_loggingConfiguration :: Lens' CreateEnvironment (Maybe LoggingConfigurationInput)

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

createEnvironment_environmentClass :: Lens' CreateEnvironment (Maybe Text)

The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

createEnvironment_tags :: Lens' CreateEnvironment (Maybe (HashMap Text Text))

The key-value tag pairs you want to associate to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

createEnvironment_requirementsS3Path :: Lens' CreateEnvironment (Maybe Text)

The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

createEnvironment_maxWorkers :: Lens' CreateEnvironment (Maybe Natural)

The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

createEnvironment_dagS3Path :: Lens' CreateEnvironment Text

The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

createEnvironment_executionRoleArn :: Lens' CreateEnvironment Text

The Amazon Resource Name (ARN) of the execution role for your environment. An execution role is an AWS Identity and Access Management (IAM) role that grants MWAA permission to access AWS services and resources used by your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

createEnvironment_name :: Lens' CreateEnvironment Text

The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

createEnvironment_networkConfiguration :: Lens' CreateEnvironment NetworkConfiguration

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

createEnvironment_sourceBucketArn :: Lens' CreateEnvironment Text

The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

Destructuring the Response

data CreateEnvironmentResponse

See: newCreateEnvironmentResponse smart constructor.

Constructors

CreateEnvironmentResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) returned in the response for the environment.

httpStatus :: Int

The response's http status code.

newCreateEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:CreateEnvironmentResponse'

-> CreateEnvironmentResponse

Create a value of CreateEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateEnvironmentResponse', createEnvironmentResponse_arn - The Amazon Resource Name (ARN) returned in the response for the environment.

$sel:httpStatus:CreateEnvironmentResponse', createEnvironmentResponse_httpStatus - The response's http status code.

Response Lenses

createEnvironmentResponse_arn :: Lens' CreateEnvironmentResponse (Maybe Text)

The Amazon Resource Name (ARN) returned in the response for the environment.

createEnvironmentResponse_httpStatus :: Lens' CreateEnvironmentResponse Int

The response's http status code.