Amazonka.MwAA.Lens

Description

Operations

ListEnvironments

listEnvironments_nextToken :: Lens' ListEnvironments (Maybe Text)

Retrieves the next page of the results.

listEnvironments_maxResults :: Lens' ListEnvironments (Maybe Natural)

The maximum number of results to retrieve per page. For example, 5 environments per page.

listEnvironmentsResponse_nextToken :: Lens' ListEnvironmentsResponse (Maybe Text)

Retrieves the next page of the results.

listEnvironmentsResponse_httpStatus :: Lens' ListEnvironmentsResponse Int

The response's http status code.

listEnvironmentsResponse_environments :: Lens' ListEnvironmentsResponse [Text]

Returns the list of Amazon MWAA environments.

UpdateEnvironment

updateEnvironment_schedulers :: Lens' UpdateEnvironment (Maybe Int)

The number of Apache Airflow schedulers to run in your Amazon MWAA environment.

updateEnvironment_minWorkers :: Lens' UpdateEnvironment (Maybe Natural)

The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

updateEnvironment_pluginsS3Path :: Lens' UpdateEnvironment (Maybe Text)

The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

updateEnvironment_webserverAccessMode :: Lens' UpdateEnvironment (Maybe WebserverAccessMode)

The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

updateEnvironment_airflowVersion :: Lens' UpdateEnvironment (Maybe Text)

The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

updateEnvironment_weeklyMaintenanceWindowStart :: Lens' UpdateEnvironment (Maybe Text)

The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following:

MON|TUE|WED|THU|FRI|SAT|SUN:([01]\\d|2[0-3]):(00|30)

updateEnvironment_executionRoleArn :: Lens' UpdateEnvironment (Maybe Text)

The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

updateEnvironment_requirementsS3ObjectVersion :: Lens' UpdateEnvironment (Maybe Text)

The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

updateEnvironment_sourceBucketArn :: Lens' UpdateEnvironment (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

updateEnvironment_dagS3Path :: Lens' UpdateEnvironment (Maybe Text)

The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

updateEnvironment_pluginsS3ObjectVersion :: Lens' UpdateEnvironment (Maybe Text)

The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

updateEnvironment_airflowConfigurationOptions :: Lens' UpdateEnvironment (Maybe (HashMap Text Text))

A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

updateEnvironment_loggingConfiguration :: Lens' UpdateEnvironment (Maybe LoggingConfigurationInput)

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

updateEnvironment_environmentClass :: Lens' UpdateEnvironment (Maybe Text)

The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

updateEnvironment_networkConfiguration :: Lens' UpdateEnvironment (Maybe UpdateNetworkConfigurationInput)

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

updateEnvironment_requirementsS3Path :: Lens' UpdateEnvironment (Maybe Text)

The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

updateEnvironment_maxWorkers :: Lens' UpdateEnvironment (Maybe Natural)

The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

updateEnvironment_name :: Lens' UpdateEnvironment Text

The name of your Amazon MWAA environment. For example, MyMWAAEnvironment.

updateEnvironmentResponse_arn :: Lens' UpdateEnvironmentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

updateEnvironmentResponse_httpStatus :: Lens' UpdateEnvironmentResponse Int

The response's http status code.

DeleteEnvironment

deleteEnvironment_name :: Lens' DeleteEnvironment Text

The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

deleteEnvironmentResponse_httpStatus :: Lens' DeleteEnvironmentResponse Int

The response's http status code.

CreateWebLoginToken

createWebLoginToken_name :: Lens' CreateWebLoginToken Text

Create an Airflow Web UI login token request for a MWAA environment.

createWebLoginTokenResponse_webServerHostname :: Lens' CreateWebLoginTokenResponse (Maybe Text)

Create an Airflow Web UI login token response for the provided webserver hostname.

createWebLoginTokenResponse_webToken :: Lens' CreateWebLoginTokenResponse (Maybe Text)

Create an Airflow Web UI login token response for the provided JWT token.

createWebLoginTokenResponse_httpStatus :: Lens' CreateWebLoginTokenResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The key-value tag pairs associated to your environment. To learn more, see Tagging AWS resources.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The key-value tag pairs you want to associate to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

GetEnvironment

getEnvironment_name :: Lens' GetEnvironment Text

The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

getEnvironmentResponse_environment :: Lens' GetEnvironmentResponse (Maybe Environment)

An object containing all available details about the environment.

getEnvironmentResponse_httpStatus :: Lens' GetEnvironmentResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

untagResource_tagKeys :: Lens' UntagResource [Text]

The key-value tag pair you want to remove. For example, "Environment": "Staging".

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

PublishMetrics

publishMetrics_environmentName :: Lens' PublishMetrics Text

Publishes environment metric data to Amazon CloudWatch.

publishMetrics_metricData :: Lens' PublishMetrics [MetricDatum]

Publishes metric data points to Amazon CloudWatch. CloudWatch associates the data points with the specified metrica.

publishMetricsResponse_httpStatus :: Lens' PublishMetricsResponse Int

The response's http status code.

CreateEnvironment

createEnvironment_schedulers :: Lens' CreateEnvironment (Maybe Int)

The number of Apache Airflow schedulers to run in your environment.

createEnvironment_minWorkers :: Lens' CreateEnvironment (Maybe Natural)

The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

createEnvironment_pluginsS3Path :: Lens' CreateEnvironment (Maybe Text)

The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

createEnvironment_webserverAccessMode :: Lens' CreateEnvironment (Maybe WebserverAccessMode)

The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

createEnvironment_airflowVersion :: Lens' CreateEnvironment (Maybe Text)

The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

createEnvironment_kmsKey :: Lens' CreateEnvironment (Maybe Text)

The AWS Key Management Service (KMS) key to encrypt the data in your environment. You can use an AWS owned CMK, or a Customer managed CMK (advanced). To learn more, see Get started with Amazon Managed Workflows for Apache Airflow.

createEnvironment_weeklyMaintenanceWindowStart :: Lens' CreateEnvironment (Maybe Text)

The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following:

MON|TUE|WED|THU|FRI|SAT|SUN:([01]\\d|2[0-3]):(00|30)

createEnvironment_requirementsS3ObjectVersion :: Lens' CreateEnvironment (Maybe Text)

The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

createEnvironment_pluginsS3ObjectVersion :: Lens' CreateEnvironment (Maybe Text)

The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

createEnvironment_airflowConfigurationOptions :: Lens' CreateEnvironment (Maybe (HashMap Text Text))

A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

createEnvironment_loggingConfiguration :: Lens' CreateEnvironment (Maybe LoggingConfigurationInput)

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

createEnvironment_environmentClass :: Lens' CreateEnvironment (Maybe Text)

The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

createEnvironment_tags :: Lens' CreateEnvironment (Maybe (HashMap Text Text))

The key-value tag pairs you want to associate to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

createEnvironment_requirementsS3Path :: Lens' CreateEnvironment (Maybe Text)

The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

createEnvironment_maxWorkers :: Lens' CreateEnvironment (Maybe Natural)

The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

createEnvironment_dagS3Path :: Lens' CreateEnvironment Text

The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

createEnvironment_executionRoleArn :: Lens' CreateEnvironment Text

The Amazon Resource Name (ARN) of the execution role for your environment. An execution role is an AWS Identity and Access Management (IAM) role that grants MWAA permission to access AWS services and resources used by your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

createEnvironment_name :: Lens' CreateEnvironment Text

The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

createEnvironment_networkConfiguration :: Lens' CreateEnvironment NetworkConfiguration

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

createEnvironment_sourceBucketArn :: Lens' CreateEnvironment Text

The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

createEnvironmentResponse_arn :: Lens' CreateEnvironmentResponse (Maybe Text)

The Amazon Resource Name (ARN) returned in the response for the environment.

createEnvironmentResponse_httpStatus :: Lens' CreateEnvironmentResponse Int

The response's http status code.

CreateCliToken

createCliToken_name :: Lens' CreateCliToken Text

Create a CLI token request for a MWAA environment.

createCliTokenResponse_webServerHostname :: Lens' CreateCliTokenResponse (Maybe Text)

Create an Airflow CLI login token response for the provided webserver hostname.

createCliTokenResponse_cliToken :: Lens' CreateCliTokenResponse (Maybe Text)

Create an Airflow CLI login token response for the provided JWT token.

createCliTokenResponse_httpStatus :: Lens' CreateCliTokenResponse Int

The response's http status code.

Types

Dimension

dimension_name :: Lens' Dimension Text

Internal only API.

dimension_value :: Lens' Dimension Text

Internal only API.

Environment

environment_serviceRoleArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) for the service-linked role of the environment. To learn more, see Amazon MWAA Service-linked role.

environment_schedulers :: Lens' Environment (Maybe Int)

The number of Apache Airflow schedulers that run in your Amazon MWAA environment.

environment_status :: Lens' Environment (Maybe EnvironmentStatus)

The status of the Amazon MWAA environment. Valid values:

CREATING - Indicates the request to create the environment is in progress.

CREATE_FAILED - Indicates the request to create the environment failed, and the environment could not be created.

AVAILABLE - Indicates the request was successful and the environment is ready to use.

UPDATING - Indicates the request to update the environment is in progress.

DELETING - Indicates the request to delete the environment is in progress.

DELETED - Indicates the request to delete the environment is complete, and the environment has been deleted.

UNAVAILABLE - Indicates the request failed, but the environment was unable to rollback and is not in a stable state.

UPDATE_FAILED - Indicates the request to update the environment failed, and the environment has rolled back successfully and is ready to use.

We recommend reviewing our troubleshooting guide for a list of common errors and their solutions. To learn more, see Amazon MWAA troubleshooting.

environment_minWorkers :: Lens' Environment (Maybe Natural)

The minimum number of workers that run in your environment. For example, 2.

environment_pluginsS3Path :: Lens' Environment (Maybe Text)

The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. To learn more, see Installing custom plugins.

environment_webserverAccessMode :: Lens' Environment (Maybe WebserverAccessMode)

The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

environment_airflowVersion :: Lens' Environment (Maybe Text)

The Apache Airflow version on your environment. For example, v1.10.12.

environment_kmsKey :: Lens' Environment (Maybe Text)

The Key Management Service (KMS) encryption key used to encrypt the data in your environment.

environment_arn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon MWAA environment.

environment_createdAt :: Lens' Environment (Maybe UTCTime)

The day and time the environment was created.

environment_weeklyMaintenanceWindowStart :: Lens' Environment (Maybe Text)

The day and time of the week that weekly maintenance updates are scheduled. For example: TUE:03:30.

environment_executionRoleArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

environment_requirementsS3ObjectVersion :: Lens' Environment (Maybe Text)

The version of the requirements.txt file on your Amazon S3 bucket. To learn more, see Installing Python dependencies.

environment_lastUpdate :: Lens' Environment (Maybe LastUpdate)

Undocumented member.

environment_sourceBucketArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

environment_webserverUrl :: Lens' Environment (Maybe Text)

The Apache Airflow Web server host name for the Amazon MWAA environment. To learn more, see Accessing the Apache Airflow UI.

environment_dagS3Path :: Lens' Environment (Maybe Text)

The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

environment_name :: Lens' Environment (Maybe Text)

The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

environment_pluginsS3ObjectVersion :: Lens' Environment (Maybe Text)

The version of the plugins.zip file on your Amazon S3 bucket. To learn more, see Installing custom plugins.

environment_airflowConfigurationOptions :: Lens' Environment (Maybe (HashMap Text Text))

A list of key-value pairs containing the Apache Airflow configuration options attached to your environment. To learn more, see Apache Airflow configuration options.

environment_loggingConfiguration :: Lens' Environment (Maybe LoggingConfiguration)

The Apache Airflow logs being sent to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

environment_environmentClass :: Lens' Environment (Maybe Text)

The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

environment_networkConfiguration :: Lens' Environment (Maybe NetworkConfiguration)

Undocumented member.

environment_tags :: Lens' Environment (Maybe (HashMap Text Text))

The key-value tag pairs associated to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

environment_requirementsS3Path :: Lens' Environment (Maybe Text)

The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. To learn more, see Installing Python dependencies.

environment_maxWorkers :: Lens' Environment (Maybe Natural)

The maximum number of workers that run in your environment. For example, 20.

LastUpdate

lastUpdate_status :: Lens' LastUpdate (Maybe UpdateStatus)

The status of the last update on the environment. Valid values: SUCCESS, PENDING, FAILED.

lastUpdate_createdAt :: Lens' LastUpdate (Maybe UTCTime)

The day and time of the last update on the environment.

lastUpdate_error :: Lens' LastUpdate (Maybe UpdateError)

The error that was encountered during the last update of the environment.

LoggingConfiguration

loggingConfiguration_taskLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

loggingConfiguration_webserverLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

loggingConfiguration_schedulerLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

loggingConfiguration_dagProcessingLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

loggingConfiguration_workerLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

LoggingConfigurationInput

loggingConfigurationInput_taskLogs :: Lens' LoggingConfigurationInput (Maybe ModuleLoggingConfigurationInput)

Undocumented member.

loggingConfigurationInput_webserverLogs :: Lens' LoggingConfigurationInput (Maybe ModuleLoggingConfigurationInput)

Undocumented member.

loggingConfigurationInput_schedulerLogs :: Lens' LoggingConfigurationInput (Maybe ModuleLoggingConfigurationInput)

Undocumented member.

loggingConfigurationInput_dagProcessingLogs :: Lens' LoggingConfigurationInput (Maybe ModuleLoggingConfigurationInput)

Undocumented member.

loggingConfigurationInput_workerLogs :: Lens' LoggingConfigurationInput (Maybe ModuleLoggingConfigurationInput)

Undocumented member.

MetricDatum

metricDatum_value :: Lens' MetricDatum (Maybe Double)

Internal only API.

metricDatum_dimensions :: Lens' MetricDatum (Maybe [Dimension])

Internal only API.

metricDatum_unit :: Lens' MetricDatum (Maybe Unit)

Undocumented member.

metricDatum_statisticValues :: Lens' MetricDatum (Maybe StatisticSet)

Internal only API.

metricDatum_metricName :: Lens' MetricDatum Text

Internal only API.

metricDatum_timestamp :: Lens' MetricDatum UTCTime

Internal only API.

ModuleLoggingConfiguration

moduleLoggingConfiguration_logLevel :: Lens' ModuleLoggingConfiguration (Maybe LoggingLevel)

Defines the Apache Airflow logs to send for the log type (e.g. DagProcessingLogs) to CloudWatch Logs. Valid values: CRITICAL, ERROR, WARNING, INFO.

moduleLoggingConfiguration_enabled :: Lens' ModuleLoggingConfiguration (Maybe Bool)

Indicates whether to enable the Apache Airflow log type (e.g. DagProcessingLogs) in CloudWatch Logs.

moduleLoggingConfiguration_cloudWatchLogGroupArn :: Lens' ModuleLoggingConfiguration (Maybe Text)

The Amazon Resource Name (ARN) for the CloudWatch Logs group where the Apache Airflow log type (e.g. DagProcessingLogs) is published. For example, arn:aws:logs:us-east-1:123456789012:log-group:airflow-MyMWAAEnvironment-MwaaEnvironment-DAGProcessing:*.

ModuleLoggingConfigurationInput

moduleLoggingConfigurationInput_enabled :: Lens' ModuleLoggingConfigurationInput Bool

Indicates whether to enable the Apache Airflow log type (e.g. DagProcessingLogs) in CloudWatch Logs.

moduleLoggingConfigurationInput_logLevel :: Lens' ModuleLoggingConfigurationInput LoggingLevel

Defines the Apache Airflow logs to send for the log type (e.g. DagProcessingLogs) to CloudWatch Logs. Valid values: CRITICAL, ERROR, WARNING, INFO.

NetworkConfiguration

networkConfiguration_securityGroupIds :: Lens' NetworkConfiguration (Maybe (NonEmpty Text))

A list of 1 or more security group IDs. Accepts up to 5 security group IDs. A security group must be attached to the same VPC as the subnets. To learn more, see Security in your VPC on Amazon MWAA.

networkConfiguration_subnetIds :: Lens' NetworkConfiguration (Maybe (NonEmpty Text))

A list of 2 subnet IDs. Required to create an environment. Must be private subnets in two different availability zones. A subnet must be attached to the same VPC as the security group.

StatisticSet

statisticSet_sampleCount :: Lens' StatisticSet (Maybe Int)

Internal only API.

statisticSet_maximum :: Lens' StatisticSet (Maybe Double)

Internal only API.

statisticSet_minimum :: Lens' StatisticSet (Maybe Double)

Internal only API.

statisticSet_sum :: Lens' StatisticSet (Maybe Double)

Internal only API.

UpdateError

updateError_errorCode :: Lens' UpdateError (Maybe Text)

The error code that corresponds to the error with the last update.

updateError_errorMessage :: Lens' UpdateError (Maybe Text)

The error message that corresponds to the error code.

UpdateNetworkConfigurationInput

updateNetworkConfigurationInput_securityGroupIds :: Lens' UpdateNetworkConfigurationInput (NonEmpty Text)

A list of 1 or more security group IDs. Accepts up to 5 security group IDs. A security group must be attached to the same VPC as the subnets. To learn more, see Security in your VPC on Amazon MWAA.