Amazonka.MwAA.Types

Description

Service Configuration

defaultService :: Service

API version 2020-07-01 of the Amazon MWAA SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

ValidationException: The provided input is not valid.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Access to the Airflow Web UI or CLI has been Denied. Please follow the MWAA user guide to setup permissions to access the Web UI and CLI functionality.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

InternalServerException: An internal error has occurred.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

ResourceNotFoundException: The resource is not available.

EnvironmentStatus

newtype EnvironmentStatus

Constructors

EnvironmentStatus'

Fields

fromEnvironmentStatus :: Text

Bundled Patterns

pattern EnvironmentStatus_AVAILABLE :: EnvironmentStatus

pattern EnvironmentStatus_CREATE_FAILED :: EnvironmentStatus

pattern EnvironmentStatus_CREATING :: EnvironmentStatus

pattern EnvironmentStatus_DELETED :: EnvironmentStatus

pattern EnvironmentStatus_DELETING :: EnvironmentStatus

pattern EnvironmentStatus_UNAVAILABLE :: EnvironmentStatus

pattern EnvironmentStatus_UPDATE_FAILED :: EnvironmentStatus

pattern EnvironmentStatus_UPDATING :: EnvironmentStatus

LoggingLevel

newtype LoggingLevel

Constructors

LoggingLevel'

Fields

fromLoggingLevel :: Text

Bundled Patterns

pattern LoggingLevel_CRITICAL :: LoggingLevel

pattern LoggingLevel_DEBUG :: LoggingLevel

pattern LoggingLevel_ERROR :: LoggingLevel

pattern LoggingLevel_INFO :: LoggingLevel

pattern LoggingLevel_WARNING :: LoggingLevel

Unit

newtype Unit

Unit

Constructors

Unit'

Fields

fromUnit :: Text

Bundled Patterns

pattern Unit_Bits :: Unit

pattern Unit_Bits_Second :: Unit

pattern Unit_Bytes :: Unit

pattern Unit_Bytes_Second :: Unit

pattern Unit_Count :: Unit

pattern Unit_Count_Second :: Unit

pattern Unit_Gigabits :: Unit

pattern Unit_Gigabits_Second :: Unit

pattern Unit_Gigabytes :: Unit

pattern Unit_Gigabytes_Second :: Unit

pattern Unit_Kilobits :: Unit

pattern Unit_Kilobits_Second :: Unit

pattern Unit_Kilobytes :: Unit

pattern Unit_Kilobytes_Second :: Unit

pattern Unit_Megabits :: Unit

pattern Unit_Megabits_Second :: Unit

pattern Unit_Megabytes :: Unit

pattern Unit_Megabytes_Second :: Unit

pattern Unit_Microseconds :: Unit

pattern Unit_Milliseconds :: Unit

pattern Unit_None :: Unit

pattern Unit_Percent :: Unit

pattern Unit_Seconds :: Unit

pattern Unit_Terabits :: Unit

pattern Unit_Terabits_Second :: Unit

pattern Unit_Terabytes :: Unit

pattern Unit_Terabytes_Second :: Unit

UpdateStatus

newtype UpdateStatus

Constructors

UpdateStatus'

Fields

fromUpdateStatus :: Text

Bundled Patterns

pattern UpdateStatus_FAILED :: UpdateStatus

pattern UpdateStatus_PENDING :: UpdateStatus

pattern UpdateStatus_SUCCESS :: UpdateStatus

WebserverAccessMode

newtype WebserverAccessMode

Constructors

WebserverAccessMode'

Fields

fromWebserverAccessMode :: Text

Bundled Patterns

pattern WebserverAccessMode_PRIVATE_ONLY :: WebserverAccessMode

pattern WebserverAccessMode_PUBLIC_ONLY :: WebserverAccessMode

Dimension

data Dimension

Internal only API.

See: newDimension smart constructor.

Constructors

Dimension'

Fields

name :: Text

Internal only API.

value :: Text

Internal only API.

newDimension

Arguments

:: Text
$sel:name:Dimension'

-> Text
$sel:value:Dimension'

-> Dimension

Create a value of Dimension with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Dimension', dimension_name - Internal only API.

$sel:value:Dimension', dimension_value - Internal only API.

dimension_name :: Lens' Dimension Text

Internal only API.

dimension_value :: Lens' Dimension Text

Internal only API.

Environment

data Environment

The Amazon Managed Workflows for Apache Airflow (MWAA) environment.

See: newEnvironment smart constructor.

Constructors

Environment'

Fields

serviceRoleArn :: Maybe Text

The Amazon Resource Name (ARN) for the service-linked role of the environment. To learn more, see Amazon MWAA Service-linked role.

schedulers :: Maybe Int

The number of Apache Airflow schedulers that run in your Amazon MWAA environment.

status :: Maybe EnvironmentStatus

The status of the Amazon MWAA environment. Valid values:

CREATING - Indicates the request to create the environment is in progress.

CREATE_FAILED - Indicates the request to create the environment failed, and the environment could not be created.

AVAILABLE - Indicates the request was successful and the environment is ready to use.

UPDATING - Indicates the request to update the environment is in progress.

DELETING - Indicates the request to delete the environment is in progress.

DELETED - Indicates the request to delete the environment is complete, and the environment has been deleted.

UNAVAILABLE - Indicates the request failed, but the environment was unable to rollback and is not in a stable state.

UPDATE_FAILED - Indicates the request to update the environment failed, and the environment has rolled back successfully and is ready to use.

We recommend reviewing our troubleshooting guide for a list of common errors and their solutions. To learn more, see Amazon MWAA troubleshooting.

minWorkers :: Maybe Natural

The minimum number of workers that run in your environment. For example, 2.

pluginsS3Path :: Maybe Text

The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. To learn more, see Installing custom plugins.

webserverAccessMode :: Maybe WebserverAccessMode

The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

airflowVersion :: Maybe Text

The Apache Airflow version on your environment. For example, v1.10.12.

kmsKey :: Maybe Text

The Key Management Service (KMS) encryption key used to encrypt the data in your environment.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon MWAA environment.

createdAt :: Maybe POSIX

The day and time the environment was created.

weeklyMaintenanceWindowStart :: Maybe Text

The day and time of the week that weekly maintenance updates are scheduled. For example: TUE:03:30.

executionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

requirementsS3ObjectVersion :: Maybe Text

The version of the requirements.txt file on your Amazon S3 bucket. To learn more, see Installing Python dependencies.

lastUpdate :: Maybe LastUpdate

sourceBucketArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

webserverUrl :: Maybe Text

The Apache Airflow Web server host name for the Amazon MWAA environment. To learn more, see Accessing the Apache Airflow UI.

dagS3Path :: Maybe Text

The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

name :: Maybe Text

The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

pluginsS3ObjectVersion :: Maybe Text

The version of the plugins.zip file on your Amazon S3 bucket. To learn more, see Installing custom plugins.

airflowConfigurationOptions :: Maybe (HashMap Text Text)

A list of key-value pairs containing the Apache Airflow configuration options attached to your environment. To learn more, see Apache Airflow configuration options.

loggingConfiguration :: Maybe LoggingConfiguration

The Apache Airflow logs being sent to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

environmentClass :: Maybe Text

The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

networkConfiguration :: Maybe NetworkConfiguration

tags :: Maybe (HashMap Text Text)

The key-value tag pairs associated to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

requirementsS3Path :: Maybe Text

The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. To learn more, see Installing Python dependencies.

maxWorkers :: Maybe Natural

The maximum number of workers that run in your environment. For example, 20.

newEnvironment :: Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceRoleArn:Environment', environment_serviceRoleArn - The Amazon Resource Name (ARN) for the service-linked role of the environment. To learn more, see Amazon MWAA Service-linked role.

$sel:schedulers:Environment', environment_schedulers - The number of Apache Airflow schedulers that run in your Amazon MWAA environment.

$sel:status:Environment', environment_status - The status of the Amazon MWAA environment. Valid values:

CREATING - Indicates the request to create the environment is in progress.

CREATE_FAILED - Indicates the request to create the environment failed, and the environment could not be created.

AVAILABLE - Indicates the request was successful and the environment is ready to use.

UPDATING - Indicates the request to update the environment is in progress.

DELETING - Indicates the request to delete the environment is in progress.

DELETED - Indicates the request to delete the environment is complete, and the environment has been deleted.

UNAVAILABLE - Indicates the request failed, but the environment was unable to rollback and is not in a stable state.

UPDATE_FAILED - Indicates the request to update the environment failed, and the environment has rolled back successfully and is ready to use.

We recommend reviewing our troubleshooting guide for a list of common errors and their solutions. To learn more, see Amazon MWAA troubleshooting.

$sel:minWorkers:Environment', environment_minWorkers - The minimum number of workers that run in your environment. For example, 2.

$sel:pluginsS3Path:Environment', environment_pluginsS3Path - The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. To learn more, see Installing custom plugins.

$sel:webserverAccessMode:Environment', environment_webserverAccessMode - The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

$sel:airflowVersion:Environment', environment_airflowVersion - The Apache Airflow version on your environment. For example, v1.10.12.

$sel:kmsKey:Environment', environment_kmsKey - The Key Management Service (KMS) encryption key used to encrypt the data in your environment.

$sel:arn:Environment', environment_arn - The Amazon Resource Name (ARN) of the Amazon MWAA environment.

$sel:createdAt:Environment', environment_createdAt - The day and time the environment was created.

$sel:weeklyMaintenanceWindowStart:Environment', environment_weeklyMaintenanceWindowStart - The day and time of the week that weekly maintenance updates are scheduled. For example: TUE:03:30.

$sel:executionRoleArn:Environment', environment_executionRoleArn - The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

$sel:requirementsS3ObjectVersion:Environment', environment_requirementsS3ObjectVersion - The version of the requirements.txt file on your Amazon S3 bucket. To learn more, see Installing Python dependencies.

$sel:lastUpdate:Environment', environment_lastUpdate - Undocumented member.

$sel:sourceBucketArn:Environment', environment_sourceBucketArn - The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

$sel:webserverUrl:Environment', environment_webserverUrl - The Apache Airflow Web server host name for the Amazon MWAA environment. To learn more, see Accessing the Apache Airflow UI.

$sel:dagS3Path:Environment', environment_dagS3Path - The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

$sel:name:Environment', environment_name - The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

$sel:pluginsS3ObjectVersion:Environment', environment_pluginsS3ObjectVersion - The version of the plugins.zip file on your Amazon S3 bucket. To learn more, see Installing custom plugins.

$sel:airflowConfigurationOptions:Environment', environment_airflowConfigurationOptions - A list of key-value pairs containing the Apache Airflow configuration options attached to your environment. To learn more, see Apache Airflow configuration options.

$sel:loggingConfiguration:Environment', environment_loggingConfiguration - The Apache Airflow logs being sent to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

$sel:environmentClass:Environment', environment_environmentClass - The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

$sel:networkConfiguration:Environment', environment_networkConfiguration - Undocumented member.

$sel:tags:Environment', environment_tags - The key-value tag pairs associated to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

$sel:requirementsS3Path:Environment', environment_requirementsS3Path - The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. To learn more, see Installing Python dependencies.

$sel:maxWorkers:Environment', environment_maxWorkers - The maximum number of workers that run in your environment. For example, 20.

environment_serviceRoleArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) for the service-linked role of the environment. To learn more, see Amazon MWAA Service-linked role.

environment_schedulers :: Lens' Environment (Maybe Int)

The number of Apache Airflow schedulers that run in your Amazon MWAA environment.

environment_status :: Lens' Environment (Maybe EnvironmentStatus)

The status of the Amazon MWAA environment. Valid values:

CREATING - Indicates the request to create the environment is in progress.

CREATE_FAILED - Indicates the request to create the environment failed, and the environment could not be created.

AVAILABLE - Indicates the request was successful and the environment is ready to use.

UPDATING - Indicates the request to update the environment is in progress.

DELETING - Indicates the request to delete the environment is in progress.

DELETED - Indicates the request to delete the environment is complete, and the environment has been deleted.

UNAVAILABLE - Indicates the request failed, but the environment was unable to rollback and is not in a stable state.

UPDATE_FAILED - Indicates the request to update the environment failed, and the environment has rolled back successfully and is ready to use.

We recommend reviewing our troubleshooting guide for a list of common errors and their solutions. To learn more, see Amazon MWAA troubleshooting.

environment_minWorkers :: Lens' Environment (Maybe Natural)

The minimum number of workers that run in your environment. For example, 2.

environment_pluginsS3Path :: Lens' Environment (Maybe Text)

The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. To learn more, see Installing custom plugins.

environment_webserverAccessMode :: Lens' Environment (Maybe WebserverAccessMode)

The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

environment_airflowVersion :: Lens' Environment (Maybe Text)

The Apache Airflow version on your environment. For example, v1.10.12.

environment_kmsKey :: Lens' Environment (Maybe Text)

The Key Management Service (KMS) encryption key used to encrypt the data in your environment.

environment_arn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon MWAA environment.

environment_createdAt :: Lens' Environment (Maybe UTCTime)

The day and time the environment was created.

environment_weeklyMaintenanceWindowStart :: Lens' Environment (Maybe Text)

The day and time of the week that weekly maintenance updates are scheduled. For example: TUE:03:30.

environment_executionRoleArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

environment_requirementsS3ObjectVersion :: Lens' Environment (Maybe Text)

The version of the requirements.txt file on your Amazon S3 bucket. To learn more, see Installing Python dependencies.

environment_lastUpdate :: Lens' Environment (Maybe LastUpdate)

Undocumented member.

environment_sourceBucketArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

environment_webserverUrl :: Lens' Environment (Maybe Text)

The Apache Airflow Web server host name for the Amazon MWAA environment. To learn more, see Accessing the Apache Airflow UI.

environment_dagS3Path :: Lens' Environment (Maybe Text)

The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

environment_name :: Lens' Environment (Maybe Text)

The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

environment_pluginsS3ObjectVersion :: Lens' Environment (Maybe Text)

The version of the plugins.zip file on your Amazon S3 bucket. To learn more, see Installing custom plugins.

environment_airflowConfigurationOptions :: Lens' Environment (Maybe (HashMap Text Text))

A list of key-value pairs containing the Apache Airflow configuration options attached to your environment. To learn more, see Apache Airflow configuration options.

environment_loggingConfiguration :: Lens' Environment (Maybe LoggingConfiguration)

The Apache Airflow logs being sent to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

environment_environmentClass :: Lens' Environment (Maybe Text)

The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

environment_networkConfiguration :: Lens' Environment (Maybe NetworkConfiguration)

Undocumented member.

environment_tags :: Lens' Environment (Maybe (HashMap Text Text))

The key-value tag pairs associated to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

environment_requirementsS3Path :: Lens' Environment (Maybe Text)

The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. To learn more, see Installing Python dependencies.

environment_maxWorkers :: Lens' Environment (Maybe Natural)

The maximum number of workers that run in your environment. For example, 20.

LastUpdate

data LastUpdate

The status of the last update on the environment, and any errors that were encountered.

See: newLastUpdate smart constructor.

Constructors

LastUpdate'

Fields

status :: Maybe UpdateStatus

The status of the last update on the environment. Valid values: SUCCESS, PENDING, FAILED.

createdAt :: Maybe POSIX

The day and time of the last update on the environment.

error :: Maybe UpdateError

The error that was encountered during the last update of the environment.

newLastUpdate :: LastUpdate

Create a value of LastUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:LastUpdate', lastUpdate_status - The status of the last update on the environment. Valid values: SUCCESS, PENDING, FAILED.

$sel:createdAt:LastUpdate', lastUpdate_createdAt - The day and time of the last update on the environment.

$sel:error:LastUpdate', lastUpdate_error - The error that was encountered during the last update of the environment.

lastUpdate_status :: Lens' LastUpdate (Maybe UpdateStatus)

The status of the last update on the environment. Valid values: SUCCESS, PENDING, FAILED.

lastUpdate_createdAt :: Lens' LastUpdate (Maybe UTCTime)

The day and time of the last update on the environment.

lastUpdate_error :: Lens' LastUpdate (Maybe UpdateError)

The error that was encountered during the last update of the environment.

LoggingConfiguration

data LoggingConfiguration

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

See: newLoggingConfiguration smart constructor.

Constructors

LoggingConfiguration'

Fields

taskLogs :: Maybe ModuleLoggingConfiguration

webserverLogs :: Maybe ModuleLoggingConfiguration

schedulerLogs :: Maybe ModuleLoggingConfiguration

dagProcessingLogs :: Maybe ModuleLoggingConfiguration

workerLogs :: Maybe ModuleLoggingConfiguration

newLoggingConfiguration :: LoggingConfiguration

Create a value of LoggingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskLogs:LoggingConfiguration', loggingConfiguration_taskLogs - Undocumented member.

$sel:webserverLogs:LoggingConfiguration', loggingConfiguration_webserverLogs - Undocumented member.

$sel:schedulerLogs:LoggingConfiguration', loggingConfiguration_schedulerLogs - Undocumented member.

$sel:dagProcessingLogs:LoggingConfiguration', loggingConfiguration_dagProcessingLogs - Undocumented member.

$sel:workerLogs:LoggingConfiguration', loggingConfiguration_workerLogs - Undocumented member.

loggingConfiguration_taskLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

loggingConfiguration_webserverLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

loggingConfiguration_schedulerLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

loggingConfiguration_dagProcessingLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

loggingConfiguration_workerLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

LoggingConfigurationInput

data LoggingConfigurationInput

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

See: newLoggingConfigurationInput smart constructor.

Constructors

LoggingConfigurationInput'

Fields

taskLogs :: Maybe ModuleLoggingConfigurationInput

webserverLogs :: Maybe ModuleLoggingConfigurationInput

schedulerLogs :: Maybe ModuleLoggingConfigurationInput

dagProcessingLogs :: Maybe ModuleLoggingConfigurationInput

workerLogs :: Maybe ModuleLoggingConfigurationInput

newLoggingConfigurationInput :: LoggingConfigurationInput

Create a value of LoggingConfigurationInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskLogs:LoggingConfigurationInput', loggingConfigurationInput_taskLogs - Undocumented member.

$sel:webserverLogs:LoggingConfigurationInput', loggingConfigurationInput_webserverLogs - Undocumented member.

$sel:schedulerLogs:LoggingConfigurationInput', loggingConfigurationInput_schedulerLogs - Undocumented member.

$sel:dagProcessingLogs:LoggingConfigurationInput', loggingConfigurationInput_dagProcessingLogs - Undocumented member.

$sel:workerLogs:LoggingConfigurationInput', loggingConfigurationInput_workerLogs - Undocumented member.

loggingConfigurationInput_taskLogs :: Lens' LoggingConfigurationInput (Maybe ModuleLoggingConfigurationInput)

Undocumented member.

loggingConfigurationInput_webserverLogs :: Lens' LoggingConfigurationInput (Maybe ModuleLoggingConfigurationInput)

Undocumented member.

loggingConfigurationInput_schedulerLogs :: Lens' LoggingConfigurationInput (Maybe ModuleLoggingConfigurationInput)

Undocumented member.

loggingConfigurationInput_dagProcessingLogs :: Lens' LoggingConfigurationInput (Maybe ModuleLoggingConfigurationInput)

Undocumented member.

loggingConfigurationInput_workerLogs :: Lens' LoggingConfigurationInput (Maybe ModuleLoggingConfigurationInput)

Undocumented member.

MetricDatum

data MetricDatum

Internal only API.

See: newMetricDatum smart constructor.

Constructors

MetricDatum'

Fields

value :: Maybe Double

Internal only API.

dimensions :: Maybe [Dimension]

Internal only API.

unit :: Maybe Unit

statisticValues :: Maybe StatisticSet

Internal only API.

metricName :: Text

Internal only API.

timestamp :: POSIX

Internal only API.

newMetricDatum

Arguments

:: Text
$sel:metricName:MetricDatum'

-> UTCTime
$sel:timestamp:MetricDatum'

-> MetricDatum

Create a value of MetricDatum with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:MetricDatum', metricDatum_value - Internal only API.

$sel:dimensions:MetricDatum', metricDatum_dimensions - Internal only API.

$sel:unit:MetricDatum', metricDatum_unit - Undocumented member.

$sel:statisticValues:MetricDatum', metricDatum_statisticValues - Internal only API.

$sel:metricName:MetricDatum', metricDatum_metricName - Internal only API.

$sel:timestamp:MetricDatum', metricDatum_timestamp - Internal only API.

metricDatum_value :: Lens' MetricDatum (Maybe Double)

Internal only API.

metricDatum_dimensions :: Lens' MetricDatum (Maybe [Dimension])

Internal only API.

metricDatum_unit :: Lens' MetricDatum (Maybe Unit)

Undocumented member.

metricDatum_statisticValues :: Lens' MetricDatum (Maybe StatisticSet)

Internal only API.

metricDatum_metricName :: Lens' MetricDatum Text

Internal only API.

metricDatum_timestamp :: Lens' MetricDatum UTCTime

Internal only API.

ModuleLoggingConfiguration

data ModuleLoggingConfiguration

Defines the type of logs to send for the Apache Airflow log type (e.g. DagProcessingLogs). Valid values: CloudWatchLogGroupArn, Enabled, LogLevel.

See: newModuleLoggingConfiguration smart constructor.

Constructors

ModuleLoggingConfiguration'

Fields

logLevel :: Maybe LoggingLevel

Defines the Apache Airflow logs to send for the log type (e.g. DagProcessingLogs) to CloudWatch Logs. Valid values: CRITICAL, ERROR, WARNING, INFO.

enabled :: Maybe Bool

Indicates whether to enable the Apache Airflow log type (e.g. DagProcessingLogs) in CloudWatch Logs.

cloudWatchLogGroupArn :: Maybe Text

The Amazon Resource Name (ARN) for the CloudWatch Logs group where the Apache Airflow log type (e.g. DagProcessingLogs) is published. For example, arn:aws:logs:us-east-1:123456789012:log-group:airflow-MyMWAAEnvironment-MwaaEnvironment-DAGProcessing:*.

newModuleLoggingConfiguration :: ModuleLoggingConfiguration

Create a value of ModuleLoggingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logLevel:ModuleLoggingConfiguration', moduleLoggingConfiguration_logLevel - Defines the Apache Airflow logs to send for the log type (e.g. DagProcessingLogs) to CloudWatch Logs. Valid values: CRITICAL, ERROR, WARNING, INFO.

$sel:enabled:ModuleLoggingConfiguration', moduleLoggingConfiguration_enabled - Indicates whether to enable the Apache Airflow log type (e.g. DagProcessingLogs) in CloudWatch Logs.

$sel:cloudWatchLogGroupArn:ModuleLoggingConfiguration', moduleLoggingConfiguration_cloudWatchLogGroupArn - The Amazon Resource Name (ARN) for the CloudWatch Logs group where the Apache Airflow log type (e.g. DagProcessingLogs) is published. For example, arn:aws:logs:us-east-1:123456789012:log-group:airflow-MyMWAAEnvironment-MwaaEnvironment-DAGProcessing:*.

moduleLoggingConfiguration_logLevel :: Lens' ModuleLoggingConfiguration (Maybe LoggingLevel)

Defines the Apache Airflow logs to send for the log type (e.g. DagProcessingLogs) to CloudWatch Logs. Valid values: CRITICAL, ERROR, WARNING, INFO.

moduleLoggingConfiguration_enabled :: Lens' ModuleLoggingConfiguration (Maybe Bool)

Indicates whether to enable the Apache Airflow log type (e.g. DagProcessingLogs) in CloudWatch Logs.

moduleLoggingConfiguration_cloudWatchLogGroupArn :: Lens' ModuleLoggingConfiguration (Maybe Text)

The Amazon Resource Name (ARN) for the CloudWatch Logs group where the Apache Airflow log type (e.g. DagProcessingLogs) is published. For example, arn:aws:logs:us-east-1:123456789012:log-group:airflow-MyMWAAEnvironment-MwaaEnvironment-DAGProcessing:*.

ModuleLoggingConfigurationInput

data ModuleLoggingConfigurationInput

Defines the type of logs to send for the Apache Airflow log type (e.g. DagProcessingLogs). Valid values: CloudWatchLogGroupArn, Enabled, LogLevel.

See: newModuleLoggingConfigurationInput smart constructor.

Constructors

ModuleLoggingConfigurationInput'

Fields

enabled :: Bool

Indicates whether to enable the Apache Airflow log type (e.g. DagProcessingLogs) in CloudWatch Logs.

logLevel :: LoggingLevel

Defines the Apache Airflow logs to send for the log type (e.g. DagProcessingLogs) to CloudWatch Logs. Valid values: CRITICAL, ERROR, WARNING, INFO.

newModuleLoggingConfigurationInput

Arguments

:: Bool
$sel:enabled:ModuleLoggingConfigurationInput'

-> LoggingLevel
$sel:logLevel:ModuleLoggingConfigurationInput'

-> ModuleLoggingConfigurationInput

Create a value of ModuleLoggingConfigurationInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:ModuleLoggingConfigurationInput', moduleLoggingConfigurationInput_enabled - Indicates whether to enable the Apache Airflow log type (e.g. DagProcessingLogs) in CloudWatch Logs.

$sel:logLevel:ModuleLoggingConfigurationInput', moduleLoggingConfigurationInput_logLevel - Defines the Apache Airflow logs to send for the log type (e.g. DagProcessingLogs) to CloudWatch Logs. Valid values: CRITICAL, ERROR, WARNING, INFO.

moduleLoggingConfigurationInput_enabled :: Lens' ModuleLoggingConfigurationInput Bool

Indicates whether to enable the Apache Airflow log type (e.g. DagProcessingLogs) in CloudWatch Logs.

moduleLoggingConfigurationInput_logLevel :: Lens' ModuleLoggingConfigurationInput LoggingLevel

Defines the Apache Airflow logs to send for the log type (e.g. DagProcessingLogs) to CloudWatch Logs. Valid values: CRITICAL, ERROR, WARNING, INFO.

NetworkConfiguration

data NetworkConfiguration

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

See: newNetworkConfiguration smart constructor.

Constructors

NetworkConfiguration'

Fields

securityGroupIds :: Maybe (NonEmpty Text)

A list of 1 or more security group IDs. Accepts up to 5 security group IDs. A security group must be attached to the same VPC as the subnets. To learn more, see Security in your VPC on Amazon MWAA.

subnetIds :: Maybe (NonEmpty Text)

A list of 2 subnet IDs. Required to create an environment. Must be private subnets in two different availability zones. A subnet must be attached to the same VPC as the security group.

newNetworkConfiguration :: NetworkConfiguration

Create a value of NetworkConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:NetworkConfiguration', networkConfiguration_securityGroupIds - A list of 1 or more security group IDs. Accepts up to 5 security group IDs. A security group must be attached to the same VPC as the subnets. To learn more, see Security in your VPC on Amazon MWAA.

$sel:subnetIds:NetworkConfiguration', networkConfiguration_subnetIds - A list of 2 subnet IDs. Required to create an environment. Must be private subnets in two different availability zones. A subnet must be attached to the same VPC as the security group.

networkConfiguration_securityGroupIds :: Lens' NetworkConfiguration (Maybe (NonEmpty Text))

A list of 1 or more security group IDs. Accepts up to 5 security group IDs. A security group must be attached to the same VPC as the subnets. To learn more, see Security in your VPC on Amazon MWAA.

networkConfiguration_subnetIds :: Lens' NetworkConfiguration (Maybe (NonEmpty Text))

A list of 2 subnet IDs. Required to create an environment. Must be private subnets in two different availability zones. A subnet must be attached to the same VPC as the security group.

StatisticSet

data StatisticSet

Internal only API.

See: newStatisticSet smart constructor.

Constructors

StatisticSet'

Fields

sampleCount :: Maybe Int

Internal only API.

maximum :: Maybe Double

Internal only API.

minimum :: Maybe Double

Internal only API.

sum :: Maybe Double

Internal only API.

newStatisticSet :: StatisticSet

Create a value of StatisticSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sampleCount:StatisticSet', statisticSet_sampleCount - Internal only API.

$sel:maximum:StatisticSet', statisticSet_maximum - Internal only API.

$sel:minimum:StatisticSet', statisticSet_minimum - Internal only API.

$sel:sum:StatisticSet', statisticSet_sum - Internal only API.

statisticSet_sampleCount :: Lens' StatisticSet (Maybe Int)

Internal only API.

statisticSet_maximum :: Lens' StatisticSet (Maybe Double)

Internal only API.

statisticSet_minimum :: Lens' StatisticSet (Maybe Double)

Internal only API.

statisticSet_sum :: Lens' StatisticSet (Maybe Double)

Internal only API.

UpdateError

data UpdateError

An object containing the error encountered with the last update: ErrorCode, ErrorMessage.

See: newUpdateError smart constructor.

Constructors

UpdateError'

Fields

errorCode :: Maybe Text

The error code that corresponds to the error with the last update.

errorMessage :: Maybe Text

The error message that corresponds to the error code.

newUpdateError :: UpdateError

Create a value of UpdateError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:UpdateError', updateError_errorCode - The error code that corresponds to the error with the last update.

$sel:errorMessage:UpdateError', updateError_errorMessage - The error message that corresponds to the error code.

updateError_errorCode :: Lens' UpdateError (Maybe Text)

The error code that corresponds to the error with the last update.

updateError_errorMessage :: Lens' UpdateError (Maybe Text)

The error message that corresponds to the error code.

UpdateNetworkConfigurationInput

data UpdateNetworkConfigurationInput

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

See: newUpdateNetworkConfigurationInput smart constructor.

Constructors

UpdateNetworkConfigurationInput'

Fields

securityGroupIds :: NonEmpty Text

A list of 1 or more security group IDs. Accepts up to 5 security group IDs. A security group must be attached to the same VPC as the subnets. To learn more, see Security in your VPC on Amazon MWAA.

newUpdateNetworkConfigurationInput

Arguments

:: NonEmpty Text
$sel:securityGroupIds:UpdateNetworkConfigurationInput'

-> UpdateNetworkConfigurationInput

Create a value of UpdateNetworkConfigurationInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:UpdateNetworkConfigurationInput', updateNetworkConfigurationInput_securityGroupIds - A list of 1 or more security group IDs. Accepts up to 5 security group IDs. A security group must be attached to the same VPC as the subnets. To learn more, see Security in your VPC on Amazon MWAA.

updateNetworkConfigurationInput_securityGroupIds :: Lens' UpdateNetworkConfigurationInput (NonEmpty Text)

A list of 1 or more security group IDs. Accepts up to 5 security group IDs. A security group must be attached to the same VPC as the subnets. To learn more, see Security in your VPC on Amazon MWAA.