Amazonka.Neptune.CopyDBClusterSnapshot

Description

Copies a snapshot of a DB cluster.

To copy a DB cluster snapshot from a shared manual DB cluster snapshot, SourceDBClusterSnapshotIdentifier must be the Amazon Resource Name (ARN) of the shared DB cluster snapshot.

Creating a Request

data CopyDBClusterSnapshot

See: newCopyDBClusterSnapshot smart constructor.

Constructors

CopyDBClusterSnapshot'

Fields

preSignedUrl :: Maybe Text

Not currently supported.

copyTags :: Maybe Bool

True to copy all tags from the source DB cluster snapshot to the target DB cluster snapshot, and otherwise false. The default is false.

kmsKeyId :: Maybe Text

The Amazon Amazon KMS key ID for an encrypted DB cluster snapshot. The KMS key ID is the Amazon Resource Name (ARN), KMS key identifier, or the KMS key alias for the KMS encryption key.

If you copy an encrypted DB cluster snapshot from your Amazon account, you can specify a value for KmsKeyId to encrypt the copy with a new KMS encryption key. If you don't specify a value for KmsKeyId, then the copy of the DB cluster snapshot is encrypted with the same KMS key as the source DB cluster snapshot.

If you copy an encrypted DB cluster snapshot that is shared from another Amazon account, then you must specify a value for KmsKeyId.

KMS encryption keys are specific to the Amazon Region that they are created in, and you can't use encryption keys from one Amazon Region in another Amazon Region.

You cannot encrypt an unencrypted DB cluster snapshot when you copy it. If you try to copy an unencrypted DB cluster snapshot and specify a value for the KmsKeyId parameter, an error is returned.

tags :: Maybe [Tag]

The tags to assign to the new DB cluster snapshot copy.

sourceDBClusterSnapshotIdentifier :: Text

The identifier of the DB cluster snapshot to copy. This parameter is not case-sensitive.

Constraints:

Must specify a valid system snapshot in the "available" state.

Specify a valid DB snapshot identifier.

Example: my-cluster-snapshot1

targetDBClusterSnapshotIdentifier :: Text

The identifier of the new DB cluster snapshot to create from the source DB cluster snapshot. This parameter is not case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster-snapshot2

newCopyDBClusterSnapshot

Arguments

:: Text
$sel:sourceDBClusterSnapshotIdentifier:CopyDBClusterSnapshot'

-> Text
$sel:targetDBClusterSnapshotIdentifier:CopyDBClusterSnapshot'

-> CopyDBClusterSnapshot

Create a value of CopyDBClusterSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preSignedUrl:CopyDBClusterSnapshot', copyDBClusterSnapshot_preSignedUrl - Not currently supported.

$sel:copyTags:CopyDBClusterSnapshot', copyDBClusterSnapshot_copyTags - True to copy all tags from the source DB cluster snapshot to the target DB cluster snapshot, and otherwise false. The default is false.

$sel:kmsKeyId:CopyDBClusterSnapshot', copyDBClusterSnapshot_kmsKeyId - The Amazon Amazon KMS key ID for an encrypted DB cluster snapshot. The KMS key ID is the Amazon Resource Name (ARN), KMS key identifier, or the KMS key alias for the KMS encryption key.

If you copy an encrypted DB cluster snapshot from your Amazon account, you can specify a value for KmsKeyId to encrypt the copy with a new KMS encryption key. If you don't specify a value for KmsKeyId, then the copy of the DB cluster snapshot is encrypted with the same KMS key as the source DB cluster snapshot.

If you copy an encrypted DB cluster snapshot that is shared from another Amazon account, then you must specify a value for KmsKeyId.

KMS encryption keys are specific to the Amazon Region that they are created in, and you can't use encryption keys from one Amazon Region in another Amazon Region.

You cannot encrypt an unencrypted DB cluster snapshot when you copy it. If you try to copy an unencrypted DB cluster snapshot and specify a value for the KmsKeyId parameter, an error is returned.

$sel:tags:CopyDBClusterSnapshot', copyDBClusterSnapshot_tags - The tags to assign to the new DB cluster snapshot copy.

$sel:sourceDBClusterSnapshotIdentifier:CopyDBClusterSnapshot', copyDBClusterSnapshot_sourceDBClusterSnapshotIdentifier - The identifier of the DB cluster snapshot to copy. This parameter is not case-sensitive.

Constraints:

Must specify a valid system snapshot in the "available" state.

Specify a valid DB snapshot identifier.

Example: my-cluster-snapshot1

$sel:targetDBClusterSnapshotIdentifier:CopyDBClusterSnapshot', copyDBClusterSnapshot_targetDBClusterSnapshotIdentifier - The identifier of the new DB cluster snapshot to create from the source DB cluster snapshot. This parameter is not case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster-snapshot2

Request Lenses

copyDBClusterSnapshot_preSignedUrl :: Lens' CopyDBClusterSnapshot (Maybe Text)

Not currently supported.

copyDBClusterSnapshot_copyTags :: Lens' CopyDBClusterSnapshot (Maybe Bool)

True to copy all tags from the source DB cluster snapshot to the target DB cluster snapshot, and otherwise false. The default is false.

copyDBClusterSnapshot_kmsKeyId :: Lens' CopyDBClusterSnapshot (Maybe Text)

The Amazon Amazon KMS key ID for an encrypted DB cluster snapshot. The KMS key ID is the Amazon Resource Name (ARN), KMS key identifier, or the KMS key alias for the KMS encryption key.

If you copy an encrypted DB cluster snapshot from your Amazon account, you can specify a value for KmsKeyId to encrypt the copy with a new KMS encryption key. If you don't specify a value for KmsKeyId, then the copy of the DB cluster snapshot is encrypted with the same KMS key as the source DB cluster snapshot.

If you copy an encrypted DB cluster snapshot that is shared from another Amazon account, then you must specify a value for KmsKeyId.

KMS encryption keys are specific to the Amazon Region that they are created in, and you can't use encryption keys from one Amazon Region in another Amazon Region.

You cannot encrypt an unencrypted DB cluster snapshot when you copy it. If you try to copy an unencrypted DB cluster snapshot and specify a value for the KmsKeyId parameter, an error is returned.

copyDBClusterSnapshot_tags :: Lens' CopyDBClusterSnapshot (Maybe [Tag])

The tags to assign to the new DB cluster snapshot copy.

copyDBClusterSnapshot_sourceDBClusterSnapshotIdentifier :: Lens' CopyDBClusterSnapshot Text

The identifier of the DB cluster snapshot to copy. This parameter is not case-sensitive.

Constraints:

Must specify a valid system snapshot in the "available" state.

Specify a valid DB snapshot identifier.

Example: my-cluster-snapshot1

copyDBClusterSnapshot_targetDBClusterSnapshotIdentifier :: Lens' CopyDBClusterSnapshot Text

The identifier of the new DB cluster snapshot to create from the source DB cluster snapshot. This parameter is not case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster-snapshot2

Destructuring the Response

data CopyDBClusterSnapshotResponse

See: newCopyDBClusterSnapshotResponse smart constructor.

Constructors

CopyDBClusterSnapshotResponse'

Fields

dbClusterSnapshot :: Maybe DBClusterSnapshot

httpStatus :: Int

The response's http status code.

newCopyDBClusterSnapshotResponse

Arguments

:: Int
$sel:httpStatus:CopyDBClusterSnapshotResponse'

-> CopyDBClusterSnapshotResponse

Create a value of CopyDBClusterSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterSnapshot:CopyDBClusterSnapshotResponse', copyDBClusterSnapshotResponse_dbClusterSnapshot - Undocumented member.

$sel:httpStatus:CopyDBClusterSnapshotResponse', copyDBClusterSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

copyDBClusterSnapshotResponse_dbClusterSnapshot :: Lens' CopyDBClusterSnapshotResponse (Maybe DBClusterSnapshot)

Undocumented member.

copyDBClusterSnapshotResponse_httpStatus :: Lens' CopyDBClusterSnapshotResponse Int

The response's http status code.