Amazonka.Neptune.CreateDBParameterGroup

Description

Creates a new DB parameter group.

A DB parameter group is initially created with the default parameters for the database engine used by the DB instance. To provide custom values for any of the parameters, you must modify the group after creating it using ModifyDBParameterGroup. Once you've created a DB parameter group, you need to associate it with your DB instance using ModifyDBInstance. When you associate a new DB parameter group with a running DB instance, you need to reboot the DB instance without failover for the new DB parameter group and associated settings to take effect.

After you create a DB parameter group, you should wait at least 5 minutes before creating your first DB instance that uses that DB parameter group as the default parameter group. This allows Amazon Neptune to fully complete the create action before the parameter group is used as the default for a new DB instance. This is especially important for parameters that are critical when creating the default database for a DB instance, such as the character set for the default database defined by the character_set_database parameter. You can use the Parameter Groups option of the Amazon Neptune console or the DescribeDBParameters command to verify that your DB parameter group has been created or modified.

Creating a Request

data CreateDBParameterGroup

See: newCreateDBParameterGroup smart constructor.

Constructors

CreateDBParameterGroup'

Fields

tags :: Maybe [Tag]

The tags to be assigned to the new DB parameter group.

dbParameterGroupName :: Text

The name of the DB parameter group.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

This value is stored as a lowercase string.

dbParameterGroupFamily :: Text

The DB parameter group family name. A DB parameter group can be associated with one and only one DB parameter group family, and can be applied only to a DB instance running a database engine and engine version compatible with that DB parameter group family.

description :: Text

The description for the DB parameter group.

newCreateDBParameterGroup

Arguments

:: Text
$sel:dbParameterGroupName:CreateDBParameterGroup'

-> Text
$sel:dbParameterGroupFamily:CreateDBParameterGroup'

-> Text
$sel:description:CreateDBParameterGroup'

-> CreateDBParameterGroup

Create a value of CreateDBParameterGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateDBParameterGroup', createDBParameterGroup_tags - The tags to be assigned to the new DB parameter group.

$sel:dbParameterGroupName:CreateDBParameterGroup', createDBParameterGroup_dbParameterGroupName - The name of the DB parameter group.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

This value is stored as a lowercase string.

$sel:dbParameterGroupFamily:CreateDBParameterGroup', createDBParameterGroup_dbParameterGroupFamily - The DB parameter group family name. A DB parameter group can be associated with one and only one DB parameter group family, and can be applied only to a DB instance running a database engine and engine version compatible with that DB parameter group family.

$sel:description:CreateDBParameterGroup', createDBParameterGroup_description - The description for the DB parameter group.

Request Lenses

createDBParameterGroup_tags :: Lens' CreateDBParameterGroup (Maybe [Tag])

The tags to be assigned to the new DB parameter group.

createDBParameterGroup_dbParameterGroupName :: Lens' CreateDBParameterGroup Text

The name of the DB parameter group.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

This value is stored as a lowercase string.

createDBParameterGroup_dbParameterGroupFamily :: Lens' CreateDBParameterGroup Text

The DB parameter group family name. A DB parameter group can be associated with one and only one DB parameter group family, and can be applied only to a DB instance running a database engine and engine version compatible with that DB parameter group family.

createDBParameterGroup_description :: Lens' CreateDBParameterGroup Text

The description for the DB parameter group.

Destructuring the Response

data CreateDBParameterGroupResponse

See: newCreateDBParameterGroupResponse smart constructor.

Constructors

CreateDBParameterGroupResponse'

Fields

dbParameterGroup :: Maybe DBParameterGroup

httpStatus :: Int

The response's http status code.

newCreateDBParameterGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateDBParameterGroupResponse'

-> CreateDBParameterGroupResponse

Create a value of CreateDBParameterGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbParameterGroup:CreateDBParameterGroupResponse', createDBParameterGroupResponse_dbParameterGroup - Undocumented member.

$sel:httpStatus:CreateDBParameterGroupResponse', createDBParameterGroupResponse_httpStatus - The response's http status code.

Response Lenses

createDBParameterGroupResponse_dbParameterGroup :: Lens' CreateDBParameterGroupResponse (Maybe DBParameterGroup)

Undocumented member.

createDBParameterGroupResponse_httpStatus :: Lens' CreateDBParameterGroupResponse Int

The response's http status code.