Amazonka.Neptune.DeleteDBClusterEndpoint

Description

Deletes a custom endpoint and removes it from an Amazon Neptune DB cluster.

Creating a Request

data DeleteDBClusterEndpoint

See: newDeleteDBClusterEndpoint smart constructor.

Constructors

DeleteDBClusterEndpoint'

Fields

dbClusterEndpointIdentifier :: Text

The identifier associated with the custom endpoint. This parameter is stored as a lowercase string.

newDeleteDBClusterEndpoint

Arguments

:: Text
$sel:dbClusterEndpointIdentifier:DeleteDBClusterEndpoint'

-> DeleteDBClusterEndpoint

Create a value of DeleteDBClusterEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterEndpointIdentifier:DeleteDBClusterEndpoint', deleteDBClusterEndpoint_dbClusterEndpointIdentifier - The identifier associated with the custom endpoint. This parameter is stored as a lowercase string.

Request Lenses

deleteDBClusterEndpoint_dbClusterEndpointIdentifier :: Lens' DeleteDBClusterEndpoint Text

The identifier associated with the custom endpoint. This parameter is stored as a lowercase string.

Destructuring the Response

data DeleteDBClusterEndpointResponse

This data type represents the information you need to connect to an Amazon Neptune DB cluster. This data type is used as a response element in the following actions:

CreateDBClusterEndpoint

DescribeDBClusterEndpoints

ModifyDBClusterEndpoint

DeleteDBClusterEndpoint

For the data structure that represents Amazon RDS DB instance endpoints, see Endpoint.

See: newDeleteDBClusterEndpointResponse smart constructor.

Constructors

DeleteDBClusterEndpointResponse'

Fields

status :: Maybe Text

The current status of the endpoint. One of: creating, available, deleting, inactive, modifying. The inactive state applies to an endpoint that cannot be used for a certain kind of cluster, such as a writer endpoint for a read-only secondary cluster in a global database.

dbClusterIdentifier :: Maybe Text

The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

dbClusterEndpointArn :: Maybe Text

The Amazon Resource Name (ARN) for the endpoint.

customEndpointType :: Maybe Text

The type associated with a custom endpoint. One of: READER, WRITER, ANY.

staticMembers :: Maybe [Text]

List of DB instance identifiers that are part of the custom endpoint group.

endpointType :: Maybe Text

The type of the endpoint. One of: READER, WRITER, CUSTOM.

dbClusterEndpointIdentifier :: Maybe Text

The identifier associated with the endpoint. This parameter is stored as a lowercase string.

endpoint :: Maybe Text

The DNS address of the endpoint.

dbClusterEndpointResourceIdentifier :: Maybe Text

A unique system-generated identifier for an endpoint. It remains the same for the whole life of the endpoint.

excludedMembers :: Maybe [Text]

List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

httpStatus :: Int

The response's http status code.

newDeleteDBClusterEndpointResponse

Arguments

:: Int
$sel:httpStatus:DeleteDBClusterEndpointResponse'

-> DeleteDBClusterEndpointResponse

Create a value of DeleteDBClusterEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DeleteDBClusterEndpointResponse', deleteDBClusterEndpointResponse_status - The current status of the endpoint. One of: creating, available, deleting, inactive, modifying. The inactive state applies to an endpoint that cannot be used for a certain kind of cluster, such as a writer endpoint for a read-only secondary cluster in a global database.

$sel:dbClusterIdentifier:DeleteDBClusterEndpointResponse', deleteDBClusterEndpointResponse_dbClusterIdentifier - The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

$sel:dbClusterEndpointArn:DeleteDBClusterEndpointResponse', deleteDBClusterEndpointResponse_dbClusterEndpointArn - The Amazon Resource Name (ARN) for the endpoint.

$sel:customEndpointType:DeleteDBClusterEndpointResponse', deleteDBClusterEndpointResponse_customEndpointType - The type associated with a custom endpoint. One of: READER, WRITER, ANY.

$sel:staticMembers:DeleteDBClusterEndpointResponse', deleteDBClusterEndpointResponse_staticMembers - List of DB instance identifiers that are part of the custom endpoint group.

$sel:endpointType:DeleteDBClusterEndpointResponse', deleteDBClusterEndpointResponse_endpointType - The type of the endpoint. One of: READER, WRITER, CUSTOM.

$sel:dbClusterEndpointIdentifier:DeleteDBClusterEndpoint', deleteDBClusterEndpointResponse_dbClusterEndpointIdentifier - The identifier associated with the endpoint. This parameter is stored as a lowercase string.

$sel:endpoint:DeleteDBClusterEndpointResponse', deleteDBClusterEndpointResponse_endpoint - The DNS address of the endpoint.

$sel:dbClusterEndpointResourceIdentifier:DeleteDBClusterEndpointResponse', deleteDBClusterEndpointResponse_dbClusterEndpointResourceIdentifier - A unique system-generated identifier for an endpoint. It remains the same for the whole life of the endpoint.

$sel:excludedMembers:DeleteDBClusterEndpointResponse', deleteDBClusterEndpointResponse_excludedMembers - List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

$sel:httpStatus:DeleteDBClusterEndpointResponse', deleteDBClusterEndpointResponse_httpStatus - The response's http status code.

Response Lenses

deleteDBClusterEndpointResponse_status :: Lens' DeleteDBClusterEndpointResponse (Maybe Text)

The current status of the endpoint. One of: creating, available, deleting, inactive, modifying. The inactive state applies to an endpoint that cannot be used for a certain kind of cluster, such as a writer endpoint for a read-only secondary cluster in a global database.

deleteDBClusterEndpointResponse_dbClusterIdentifier :: Lens' DeleteDBClusterEndpointResponse (Maybe Text)

The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

deleteDBClusterEndpointResponse_dbClusterEndpointArn :: Lens' DeleteDBClusterEndpointResponse (Maybe Text)

The Amazon Resource Name (ARN) for the endpoint.

deleteDBClusterEndpointResponse_customEndpointType :: Lens' DeleteDBClusterEndpointResponse (Maybe Text)

The type associated with a custom endpoint. One of: READER, WRITER, ANY.

deleteDBClusterEndpointResponse_staticMembers :: Lens' DeleteDBClusterEndpointResponse (Maybe [Text])

List of DB instance identifiers that are part of the custom endpoint group.

deleteDBClusterEndpointResponse_endpointType :: Lens' DeleteDBClusterEndpointResponse (Maybe Text)

The type of the endpoint. One of: READER, WRITER, CUSTOM.

deleteDBClusterEndpointResponse_dbClusterEndpointIdentifier :: Lens' DeleteDBClusterEndpointResponse (Maybe Text)

The identifier associated with the endpoint. This parameter is stored as a lowercase string.

deleteDBClusterEndpointResponse_endpoint :: Lens' DeleteDBClusterEndpointResponse (Maybe Text)

The DNS address of the endpoint.

deleteDBClusterEndpointResponse_dbClusterEndpointResourceIdentifier :: Lens' DeleteDBClusterEndpointResponse (Maybe Text)

A unique system-generated identifier for an endpoint. It remains the same for the whole life of the endpoint.

deleteDBClusterEndpointResponse_excludedMembers :: Lens' DeleteDBClusterEndpointResponse (Maybe [Text])

List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

deleteDBClusterEndpointResponse_httpStatus :: Lens' DeleteDBClusterEndpointResponse Int

The response's http status code.