Amazonka.Neptune.DeleteDBInstance

Description

The DeleteDBInstance action deletes a previously provisioned DB instance. When you delete a DB instance, all automated backups for that instance are deleted and can't be recovered. Manual DB snapshots of the DB instance to be deleted by DeleteDBInstance are not deleted.

If you request a final DB snapshot the status of the Amazon Neptune DB instance is deleting until the DB snapshot is created. The API action DescribeDBInstance is used to monitor the status of this operation. The action can't be canceled or reverted once submitted.

Note that when a DB instance is in a failure state and has a status of failed, incompatible-restore, or incompatible-network, you can only delete it when the SkipFinalSnapshot parameter is set to true.

You can't delete a DB instance if it is the only instance in the DB cluster, or if it has deletion protection enabled.

Creating a Request

data DeleteDBInstance

See: newDeleteDBInstance smart constructor.

Constructors

DeleteDBInstance'

Fields

finalDBSnapshotIdentifier :: Maybe Text

The DBSnapshotIdentifier of the new DBSnapshot created when SkipFinalSnapshot is set to false.

Specifying this parameter and also setting the SkipFinalShapshot parameter to true results in an error.

Constraints:

Must be 1 to 255 letters or numbers.

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Cannot be specified when deleting a Read Replica.

skipFinalSnapshot :: Maybe Bool

Determines whether a final DB snapshot is created before the DB instance is deleted. If true is specified, no DBSnapshot is created. If false is specified, a DB snapshot is created before the DB instance is deleted.

Note that when a DB instance is in a failure state and has a status of 'failed', 'incompatible-restore', or 'incompatible-network', it can only be deleted when the SkipFinalSnapshot parameter is set to "true".

Specify true when deleting a Read Replica.

The FinalDBSnapshotIdentifier parameter must be specified if SkipFinalSnapshot is false.

Default: false

dbInstanceIdentifier :: Text

The DB instance identifier for the DB instance to be deleted. This parameter isn't case-sensitive.

Constraints:

Must match the name of an existing DB instance.

newDeleteDBInstance

Arguments

:: Text
$sel:dbInstanceIdentifier:DeleteDBInstance'

-> DeleteDBInstance

Create a value of DeleteDBInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:finalDBSnapshotIdentifier:DeleteDBInstance', deleteDBInstance_finalDBSnapshotIdentifier - The DBSnapshotIdentifier of the new DBSnapshot created when SkipFinalSnapshot is set to false.

Specifying this parameter and also setting the SkipFinalShapshot parameter to true results in an error.

Constraints:

Must be 1 to 255 letters or numbers.

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Cannot be specified when deleting a Read Replica.

$sel:skipFinalSnapshot:DeleteDBInstance', deleteDBInstance_skipFinalSnapshot - Determines whether a final DB snapshot is created before the DB instance is deleted. If true is specified, no DBSnapshot is created. If false is specified, a DB snapshot is created before the DB instance is deleted.

Note that when a DB instance is in a failure state and has a status of 'failed', 'incompatible-restore', or 'incompatible-network', it can only be deleted when the SkipFinalSnapshot parameter is set to "true".

Specify true when deleting a Read Replica.

The FinalDBSnapshotIdentifier parameter must be specified if SkipFinalSnapshot is false.

Default: false

$sel:dbInstanceIdentifier:DeleteDBInstance', deleteDBInstance_dbInstanceIdentifier - The DB instance identifier for the DB instance to be deleted. This parameter isn't case-sensitive.

Constraints:

Must match the name of an existing DB instance.

Request Lenses

deleteDBInstance_finalDBSnapshotIdentifier :: Lens' DeleteDBInstance (Maybe Text)

The DBSnapshotIdentifier of the new DBSnapshot created when SkipFinalSnapshot is set to false.

Specifying this parameter and also setting the SkipFinalShapshot parameter to true results in an error.

Constraints:

Must be 1 to 255 letters or numbers.

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Cannot be specified when deleting a Read Replica.

deleteDBInstance_skipFinalSnapshot :: Lens' DeleteDBInstance (Maybe Bool)

Determines whether a final DB snapshot is created before the DB instance is deleted. If true is specified, no DBSnapshot is created. If false is specified, a DB snapshot is created before the DB instance is deleted.

Note that when a DB instance is in a failure state and has a status of 'failed', 'incompatible-restore', or 'incompatible-network', it can only be deleted when the SkipFinalSnapshot parameter is set to "true".

Specify true when deleting a Read Replica.

The FinalDBSnapshotIdentifier parameter must be specified if SkipFinalSnapshot is false.

Default: false

deleteDBInstance_dbInstanceIdentifier :: Lens' DeleteDBInstance Text

The DB instance identifier for the DB instance to be deleted. This parameter isn't case-sensitive.

Constraints:

Must match the name of an existing DB instance.

Destructuring the Response

data DeleteDBInstanceResponse

See: newDeleteDBInstanceResponse smart constructor.

Constructors

DeleteDBInstanceResponse'

Fields

dbInstance :: Maybe DBInstance

httpStatus :: Int

The response's http status code.

newDeleteDBInstanceResponse

Arguments

:: Int
$sel:httpStatus:DeleteDBInstanceResponse'

-> DeleteDBInstanceResponse

Create a value of DeleteDBInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstance:DeleteDBInstanceResponse', deleteDBInstanceResponse_dbInstance - Undocumented member.

$sel:httpStatus:DeleteDBInstanceResponse', deleteDBInstanceResponse_httpStatus - The response's http status code.

Response Lenses

deleteDBInstanceResponse_dbInstance :: Lens' DeleteDBInstanceResponse (Maybe DBInstance)

Undocumented member.

deleteDBInstanceResponse_httpStatus :: Lens' DeleteDBInstanceResponse Int

The response's http status code.