Amazonka.Neptune.DescribeDBEngineVersions

Description

Returns a list of the available DB engines.

This operation returns paginated results.

Creating a Request

data DescribeDBEngineVersions

See: newDescribeDBEngineVersions smart constructor.

Constructors

DescribeDBEngineVersions'

Fields

engineVersion :: Maybe Text

The database engine version to return.

Example: 5.1.49

listSupportedTimezones :: Maybe Bool

If this parameter is specified and the requested engine supports the TimeZone parameter for CreateDBInstance, the response includes a list of supported time zones for each engine version.

defaultOnly :: Maybe Bool

Indicates that only the default version of the specified engine or engine and major version combination is returned.

filters :: Maybe [Filter]

Not currently supported.

engine :: Maybe Text

The database engine to return.

dbParameterGroupFamily :: Maybe Text

The name of a specific DB parameter group family to return details for.

Constraints:

If supplied, must match an existing DBParameterGroupFamily.

listSupportedCharacterSets :: Maybe Bool

If this parameter is specified and the requested engine supports the CharacterSetName parameter for CreateDBInstance, the response includes a list of supported character sets for each engine version.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

newDescribeDBEngineVersions :: DescribeDBEngineVersions

Create a value of DescribeDBEngineVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:DescribeDBEngineVersions', describeDBEngineVersions_engineVersion - The database engine version to return.

Example: 5.1.49

$sel:listSupportedTimezones:DescribeDBEngineVersions', describeDBEngineVersions_listSupportedTimezones - If this parameter is specified and the requested engine supports the TimeZone parameter for CreateDBInstance, the response includes a list of supported time zones for each engine version.

$sel:defaultOnly:DescribeDBEngineVersions', describeDBEngineVersions_defaultOnly - Indicates that only the default version of the specified engine or engine and major version combination is returned.

$sel:filters:DescribeDBEngineVersions', describeDBEngineVersions_filters - Not currently supported.

$sel:engine:DescribeDBEngineVersions', describeDBEngineVersions_engine - The database engine to return.

$sel:dbParameterGroupFamily:DescribeDBEngineVersions', describeDBEngineVersions_dbParameterGroupFamily - The name of a specific DB parameter group family to return details for.

Constraints:

If supplied, must match an existing DBParameterGroupFamily.

$sel:listSupportedCharacterSets:DescribeDBEngineVersions', describeDBEngineVersions_listSupportedCharacterSets - If this parameter is specified and the requested engine supports the CharacterSetName parameter for CreateDBInstance, the response includes a list of supported character sets for each engine version.

$sel:marker:DescribeDBEngineVersions', describeDBEngineVersions_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeDBEngineVersions', describeDBEngineVersions_maxRecords - The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Request Lenses

describeDBEngineVersions_engineVersion :: Lens' DescribeDBEngineVersions (Maybe Text)

The database engine version to return.

Example: 5.1.49

describeDBEngineVersions_listSupportedTimezones :: Lens' DescribeDBEngineVersions (Maybe Bool)

If this parameter is specified and the requested engine supports the TimeZone parameter for CreateDBInstance, the response includes a list of supported time zones for each engine version.

describeDBEngineVersions_defaultOnly :: Lens' DescribeDBEngineVersions (Maybe Bool)

Indicates that only the default version of the specified engine or engine and major version combination is returned.

describeDBEngineVersions_filters :: Lens' DescribeDBEngineVersions (Maybe [Filter])

Not currently supported.

describeDBEngineVersions_engine :: Lens' DescribeDBEngineVersions (Maybe Text)

The database engine to return.

describeDBEngineVersions_dbParameterGroupFamily :: Lens' DescribeDBEngineVersions (Maybe Text)

The name of a specific DB parameter group family to return details for.

Constraints:

If supplied, must match an existing DBParameterGroupFamily.

describeDBEngineVersions_listSupportedCharacterSets :: Lens' DescribeDBEngineVersions (Maybe Bool)

If this parameter is specified and the requested engine supports the CharacterSetName parameter for CreateDBInstance, the response includes a list of supported character sets for each engine version.

describeDBEngineVersions_marker :: Lens' DescribeDBEngineVersions (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBEngineVersions_maxRecords :: Lens' DescribeDBEngineVersions (Maybe Int)

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Destructuring the Response

data DescribeDBEngineVersionsResponse

See: newDescribeDBEngineVersionsResponse smart constructor.

Constructors

DescribeDBEngineVersionsResponse'

Fields

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dbEngineVersions :: Maybe [DBEngineVersion]

A list of DBEngineVersion elements.

httpStatus :: Int

The response's http status code.

newDescribeDBEngineVersionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDBEngineVersionsResponse'

-> DescribeDBEngineVersionsResponse

Create a value of DescribeDBEngineVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeDBEngineVersions', describeDBEngineVersionsResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:dbEngineVersions:DescribeDBEngineVersionsResponse', describeDBEngineVersionsResponse_dbEngineVersions - A list of DBEngineVersion elements.

$sel:httpStatus:DescribeDBEngineVersionsResponse', describeDBEngineVersionsResponse_httpStatus - The response's http status code.

Response Lenses

describeDBEngineVersionsResponse_marker :: Lens' DescribeDBEngineVersionsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBEngineVersionsResponse_dbEngineVersions :: Lens' DescribeDBEngineVersionsResponse (Maybe [DBEngineVersion])

A list of DBEngineVersion elements.

describeDBEngineVersionsResponse_httpStatus :: Lens' DescribeDBEngineVersionsResponse Int

The response's http status code.