Amazonka.Neptune.ModifyDBCluster

Description

Modify a setting for a DB cluster. You can change one or more database configuration parameters by specifying these parameters and the new values in the request.

Creating a Request

data ModifyDBCluster

See: newModifyDBCluster smart constructor.

Constructors

ModifyDBCluster'

Fields

engineVersion :: Maybe Text

The version number of the database engine to which you want to upgrade. Changing this parameter results in an outage. The change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true.

For a list of valid engine versions, see Engine Releases for Amazon Neptune, or call DescribeDBEngineVersions.

deletionProtection :: Maybe Bool

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

masterUserPassword :: Maybe Text

Not supported by Neptune.

cloudwatchLogsExportConfiguration :: Maybe CloudwatchLogsExportConfiguration

The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB cluster.

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

backupRetentionPeriod :: Maybe Int

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

vpcSecurityGroupIds :: Maybe [Text]

A list of VPC security groups that the DB cluster will belong to.

dbClusterParameterGroupName :: Maybe Text

The name of the DB cluster parameter group to use for the DB cluster.

applyImmediately :: Maybe Bool

A value that specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB cluster. If this parameter is set to false, changes to the DB cluster are applied during the next maintenance window.

The ApplyImmediately parameter only affects NewDBClusterIdentifier values. If you set the ApplyImmediately parameter value to false, then changes to NewDBClusterIdentifier values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter.

Default: false

optionGroupName :: Maybe Text

Not supported by Neptune.

copyTagsToSnapshot :: Maybe Bool

/If set to true, tags are copied to any snapshot of the DB cluster that is created./

newDBClusterIdentifier' :: Maybe Text

The new DB cluster identifier for the DB cluster when renaming a DB cluster. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

The first character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-cluster2

port :: Maybe Int

The port number on which the DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

enableIAMDatabaseAuthentication :: Maybe Bool

True to enable mapping of Amazon Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: false

dbClusterIdentifier :: Text

The DB cluster identifier for the cluster being modified. This parameter is not case-sensitive.

Constraints:

Must match the identifier of an existing DBCluster.

newModifyDBCluster

Arguments

:: Text
$sel:dbClusterIdentifier:ModifyDBCluster'

-> ModifyDBCluster

Create a value of ModifyDBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:ModifyDBCluster', modifyDBCluster_engineVersion - The version number of the database engine to which you want to upgrade. Changing this parameter results in an outage. The change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true.

For a list of valid engine versions, see Engine Releases for Amazon Neptune, or call DescribeDBEngineVersions.

$sel:deletionProtection:ModifyDBCluster', modifyDBCluster_deletionProtection - A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

$sel:masterUserPassword:ModifyDBCluster', modifyDBCluster_masterUserPassword - Not supported by Neptune.

$sel:cloudwatchLogsExportConfiguration:ModifyDBCluster', modifyDBCluster_cloudwatchLogsExportConfiguration - The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB cluster.

$sel:preferredMaintenanceWindow:ModifyDBCluster', modifyDBCluster_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

$sel:preferredBackupWindow:ModifyDBCluster', modifyDBCluster_preferredBackupWindow - The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

$sel:backupRetentionPeriod:ModifyDBCluster', modifyDBCluster_backupRetentionPeriod - The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

$sel:vpcSecurityGroupIds:ModifyDBCluster', modifyDBCluster_vpcSecurityGroupIds - A list of VPC security groups that the DB cluster will belong to.

$sel:dbClusterParameterGroupName:ModifyDBCluster', modifyDBCluster_dbClusterParameterGroupName - The name of the DB cluster parameter group to use for the DB cluster.

$sel:applyImmediately:ModifyDBCluster', modifyDBCluster_applyImmediately - A value that specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB cluster. If this parameter is set to false, changes to the DB cluster are applied during the next maintenance window.

The ApplyImmediately parameter only affects NewDBClusterIdentifier values. If you set the ApplyImmediately parameter value to false, then changes to NewDBClusterIdentifier values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter.

Default: false

$sel:optionGroupName:ModifyDBCluster', modifyDBCluster_optionGroupName - Not supported by Neptune.

$sel:copyTagsToSnapshot:ModifyDBCluster', modifyDBCluster_copyTagsToSnapshot - /If set to true, tags are copied to any snapshot of the DB cluster that is created./

$sel:newDBClusterIdentifier':ModifyDBCluster', modifyDBCluster_newDBClusterIdentifier - The new DB cluster identifier for the DB cluster when renaming a DB cluster. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

The first character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-cluster2

$sel:port:ModifyDBCluster', modifyDBCluster_port - The port number on which the DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

$sel:enableIAMDatabaseAuthentication:ModifyDBCluster', modifyDBCluster_enableIAMDatabaseAuthentication - True to enable mapping of Amazon Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: false

$sel:dbClusterIdentifier:ModifyDBCluster', modifyDBCluster_dbClusterIdentifier - The DB cluster identifier for the cluster being modified. This parameter is not case-sensitive.

Constraints:

Must match the identifier of an existing DBCluster.

Request Lenses

modifyDBCluster_engineVersion :: Lens' ModifyDBCluster (Maybe Text)

The version number of the database engine to which you want to upgrade. Changing this parameter results in an outage. The change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true.

For a list of valid engine versions, see Engine Releases for Amazon Neptune, or call DescribeDBEngineVersions.

modifyDBCluster_deletionProtection :: Lens' ModifyDBCluster (Maybe Bool)

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

modifyDBCluster_masterUserPassword :: Lens' ModifyDBCluster (Maybe Text)

Not supported by Neptune.

modifyDBCluster_cloudwatchLogsExportConfiguration :: Lens' ModifyDBCluster (Maybe CloudwatchLogsExportConfiguration)

The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB cluster.

modifyDBCluster_preferredMaintenanceWindow :: Lens' ModifyDBCluster (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

modifyDBCluster_preferredBackupWindow :: Lens' ModifyDBCluster (Maybe Text)

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

modifyDBCluster_backupRetentionPeriod :: Lens' ModifyDBCluster (Maybe Int)

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

modifyDBCluster_vpcSecurityGroupIds :: Lens' ModifyDBCluster (Maybe [Text])

A list of VPC security groups that the DB cluster will belong to.

modifyDBCluster_dbClusterParameterGroupName :: Lens' ModifyDBCluster (Maybe Text)

The name of the DB cluster parameter group to use for the DB cluster.

modifyDBCluster_applyImmediately :: Lens' ModifyDBCluster (Maybe Bool)

A value that specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB cluster. If this parameter is set to false, changes to the DB cluster are applied during the next maintenance window.

The ApplyImmediately parameter only affects NewDBClusterIdentifier values. If you set the ApplyImmediately parameter value to false, then changes to NewDBClusterIdentifier values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter.

Default: false

modifyDBCluster_optionGroupName :: Lens' ModifyDBCluster (Maybe Text)

Not supported by Neptune.

modifyDBCluster_copyTagsToSnapshot :: Lens' ModifyDBCluster (Maybe Bool)

/If set to true, tags are copied to any snapshot of the DB cluster that is created./

modifyDBCluster_newDBClusterIdentifier :: Lens' ModifyDBCluster (Maybe Text)

The new DB cluster identifier for the DB cluster when renaming a DB cluster. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

The first character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-cluster2

modifyDBCluster_port :: Lens' ModifyDBCluster (Maybe Int)

The port number on which the DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

modifyDBCluster_enableIAMDatabaseAuthentication :: Lens' ModifyDBCluster (Maybe Bool)

True to enable mapping of Amazon Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: false

modifyDBCluster_dbClusterIdentifier :: Lens' ModifyDBCluster Text

The DB cluster identifier for the cluster being modified. This parameter is not case-sensitive.

Constraints:

Must match the identifier of an existing DBCluster.

Destructuring the Response

data ModifyDBClusterResponse

See: newModifyDBClusterResponse smart constructor.

Constructors

ModifyDBClusterResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newModifyDBClusterResponse

Arguments

:: Int
$sel:httpStatus:ModifyDBClusterResponse'

-> ModifyDBClusterResponse

Create a value of ModifyDBClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:ModifyDBClusterResponse', modifyDBClusterResponse_dbCluster - Undocumented member.

$sel:httpStatus:ModifyDBClusterResponse', modifyDBClusterResponse_httpStatus - The response's http status code.

Response Lenses

modifyDBClusterResponse_dbCluster :: Lens' ModifyDBClusterResponse (Maybe DBCluster)

Undocumented member.

modifyDBClusterResponse_httpStatus :: Lens' ModifyDBClusterResponse Int

The response's http status code.