Amazonka.Neptune.ModifyDBClusterEndpoint

Description

Modifies the properties of an endpoint in an Amazon Neptune DB cluster.

Creating a Request

data ModifyDBClusterEndpoint

See: newModifyDBClusterEndpoint smart constructor.

Constructors

ModifyDBClusterEndpoint'

Fields

staticMembers :: Maybe [Text]

List of DB instance identifiers that are part of the custom endpoint group.

endpointType :: Maybe Text

The type of the endpoint. One of: READER, WRITER, ANY.

excludedMembers :: Maybe [Text]

List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

dbClusterEndpointIdentifier :: Text

The identifier of the endpoint to modify. This parameter is stored as a lowercase string.

newModifyDBClusterEndpoint

Arguments

:: Text
$sel:dbClusterEndpointIdentifier:ModifyDBClusterEndpoint'

-> ModifyDBClusterEndpoint

Create a value of ModifyDBClusterEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:staticMembers:ModifyDBClusterEndpoint', modifyDBClusterEndpoint_staticMembers - List of DB instance identifiers that are part of the custom endpoint group.

$sel:endpointType:ModifyDBClusterEndpoint', modifyDBClusterEndpoint_endpointType - The type of the endpoint. One of: READER, WRITER, ANY.

$sel:excludedMembers:ModifyDBClusterEndpoint', modifyDBClusterEndpoint_excludedMembers - List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

$sel:dbClusterEndpointIdentifier:ModifyDBClusterEndpoint', modifyDBClusterEndpoint_dbClusterEndpointIdentifier - The identifier of the endpoint to modify. This parameter is stored as a lowercase string.

Request Lenses

modifyDBClusterEndpoint_staticMembers :: Lens' ModifyDBClusterEndpoint (Maybe [Text])

List of DB instance identifiers that are part of the custom endpoint group.

modifyDBClusterEndpoint_endpointType :: Lens' ModifyDBClusterEndpoint (Maybe Text)

The type of the endpoint. One of: READER, WRITER, ANY.

modifyDBClusterEndpoint_excludedMembers :: Lens' ModifyDBClusterEndpoint (Maybe [Text])

List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

modifyDBClusterEndpoint_dbClusterEndpointIdentifier :: Lens' ModifyDBClusterEndpoint Text

The identifier of the endpoint to modify. This parameter is stored as a lowercase string.

Destructuring the Response

data ModifyDBClusterEndpointResponse

This data type represents the information you need to connect to an Amazon Aurora DB cluster. This data type is used as a response element in the following actions:

CreateDBClusterEndpoint

DescribeDBClusterEndpoints

ModifyDBClusterEndpoint

DeleteDBClusterEndpoint

For the data structure that represents Amazon RDS DB instance endpoints, see Endpoint.

See: newModifyDBClusterEndpointResponse smart constructor.

Constructors

ModifyDBClusterEndpointResponse'

Fields

status :: Maybe Text

The current status of the endpoint. One of: creating, available, deleting, inactive, modifying. The inactive state applies to an endpoint that cannot be used for a certain kind of cluster, such as a writer endpoint for a read-only secondary cluster in a global database.

dbClusterIdentifier :: Maybe Text

The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

dbClusterEndpointArn :: Maybe Text

The Amazon Resource Name (ARN) for the endpoint.

customEndpointType :: Maybe Text

The type associated with a custom endpoint. One of: READER, WRITER, ANY.

staticMembers :: Maybe [Text]

List of DB instance identifiers that are part of the custom endpoint group.

endpointType :: Maybe Text

The type of the endpoint. One of: READER, WRITER, CUSTOM.

dbClusterEndpointIdentifier :: Maybe Text

The identifier associated with the endpoint. This parameter is stored as a lowercase string.

endpoint :: Maybe Text

The DNS address of the endpoint.

dbClusterEndpointResourceIdentifier :: Maybe Text

A unique system-generated identifier for an endpoint. It remains the same for the whole life of the endpoint.

excludedMembers :: Maybe [Text]

List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

httpStatus :: Int

The response's http status code.

newModifyDBClusterEndpointResponse

Arguments

:: Int
$sel:httpStatus:ModifyDBClusterEndpointResponse'

-> ModifyDBClusterEndpointResponse

Create a value of ModifyDBClusterEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ModifyDBClusterEndpointResponse', modifyDBClusterEndpointResponse_status - The current status of the endpoint. One of: creating, available, deleting, inactive, modifying. The inactive state applies to an endpoint that cannot be used for a certain kind of cluster, such as a writer endpoint for a read-only secondary cluster in a global database.

$sel:dbClusterIdentifier:ModifyDBClusterEndpointResponse', modifyDBClusterEndpointResponse_dbClusterIdentifier - The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

$sel:dbClusterEndpointArn:ModifyDBClusterEndpointResponse', modifyDBClusterEndpointResponse_dbClusterEndpointArn - The Amazon Resource Name (ARN) for the endpoint.

$sel:customEndpointType:ModifyDBClusterEndpointResponse', modifyDBClusterEndpointResponse_customEndpointType - The type associated with a custom endpoint. One of: READER, WRITER, ANY.

$sel:staticMembers:ModifyDBClusterEndpoint', modifyDBClusterEndpointResponse_staticMembers - List of DB instance identifiers that are part of the custom endpoint group.

$sel:endpointType:ModifyDBClusterEndpoint', modifyDBClusterEndpointResponse_endpointType - The type of the endpoint. One of: READER, WRITER, CUSTOM.

$sel:dbClusterEndpointIdentifier:ModifyDBClusterEndpoint', modifyDBClusterEndpointResponse_dbClusterEndpointIdentifier - The identifier associated with the endpoint. This parameter is stored as a lowercase string.

$sel:endpoint:ModifyDBClusterEndpointResponse', modifyDBClusterEndpointResponse_endpoint - The DNS address of the endpoint.

$sel:dbClusterEndpointResourceIdentifier:ModifyDBClusterEndpointResponse', modifyDBClusterEndpointResponse_dbClusterEndpointResourceIdentifier - A unique system-generated identifier for an endpoint. It remains the same for the whole life of the endpoint.

$sel:excludedMembers:ModifyDBClusterEndpoint', modifyDBClusterEndpointResponse_excludedMembers - List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

$sel:httpStatus:ModifyDBClusterEndpointResponse', modifyDBClusterEndpointResponse_httpStatus - The response's http status code.

Response Lenses

modifyDBClusterEndpointResponse_status :: Lens' ModifyDBClusterEndpointResponse (Maybe Text)

The current status of the endpoint. One of: creating, available, deleting, inactive, modifying. The inactive state applies to an endpoint that cannot be used for a certain kind of cluster, such as a writer endpoint for a read-only secondary cluster in a global database.

modifyDBClusterEndpointResponse_dbClusterIdentifier :: Lens' ModifyDBClusterEndpointResponse (Maybe Text)

The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

modifyDBClusterEndpointResponse_dbClusterEndpointArn :: Lens' ModifyDBClusterEndpointResponse (Maybe Text)

The Amazon Resource Name (ARN) for the endpoint.

modifyDBClusterEndpointResponse_customEndpointType :: Lens' ModifyDBClusterEndpointResponse (Maybe Text)

The type associated with a custom endpoint. One of: READER, WRITER, ANY.

modifyDBClusterEndpointResponse_staticMembers :: Lens' ModifyDBClusterEndpointResponse (Maybe [Text])

List of DB instance identifiers that are part of the custom endpoint group.

modifyDBClusterEndpointResponse_endpointType :: Lens' ModifyDBClusterEndpointResponse (Maybe Text)

The type of the endpoint. One of: READER, WRITER, CUSTOM.

modifyDBClusterEndpointResponse_dbClusterEndpointIdentifier :: Lens' ModifyDBClusterEndpointResponse (Maybe Text)

The identifier associated with the endpoint. This parameter is stored as a lowercase string.

modifyDBClusterEndpointResponse_endpoint :: Lens' ModifyDBClusterEndpointResponse (Maybe Text)

The DNS address of the endpoint.

modifyDBClusterEndpointResponse_dbClusterEndpointResourceIdentifier :: Lens' ModifyDBClusterEndpointResponse (Maybe Text)

A unique system-generated identifier for an endpoint. It remains the same for the whole life of the endpoint.

modifyDBClusterEndpointResponse_excludedMembers :: Lens' ModifyDBClusterEndpointResponse (Maybe [Text])

List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

modifyDBClusterEndpointResponse_httpStatus :: Lens' ModifyDBClusterEndpointResponse Int

The response's http status code.