Amazonka.Neptune.RemoveSourceIdentifierFromSubscription

Description

Removes a source identifier from an existing event notification subscription.

Creating a Request

data RemoveSourceIdentifierFromSubscription

See: newRemoveSourceIdentifierFromSubscription smart constructor.

Constructors

RemoveSourceIdentifierFromSubscription'

Fields

subscriptionName :: Text

The name of the event notification subscription you want to remove a source identifier from.

sourceIdentifier :: Text

The source identifier to be removed from the subscription, such as the DB instance identifier for a DB instance or the name of a security group.

newRemoveSourceIdentifierFromSubscription

Arguments

:: Text
$sel:subscriptionName:RemoveSourceIdentifierFromSubscription'

-> Text
$sel:sourceIdentifier:RemoveSourceIdentifierFromSubscription'

-> RemoveSourceIdentifierFromSubscription

Create a value of RemoveSourceIdentifierFromSubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionName:RemoveSourceIdentifierFromSubscription', removeSourceIdentifierFromSubscription_subscriptionName - The name of the event notification subscription you want to remove a source identifier from.

$sel:sourceIdentifier:RemoveSourceIdentifierFromSubscription', removeSourceIdentifierFromSubscription_sourceIdentifier - The source identifier to be removed from the subscription, such as the DB instance identifier for a DB instance or the name of a security group.

Request Lenses

removeSourceIdentifierFromSubscription_subscriptionName :: Lens' RemoveSourceIdentifierFromSubscription Text

The name of the event notification subscription you want to remove a source identifier from.

removeSourceIdentifierFromSubscription_sourceIdentifier :: Lens' RemoveSourceIdentifierFromSubscription Text

The source identifier to be removed from the subscription, such as the DB instance identifier for a DB instance or the name of a security group.

Destructuring the Response

data RemoveSourceIdentifierFromSubscriptionResponse

See: newRemoveSourceIdentifierFromSubscriptionResponse smart constructor.

Constructors

RemoveSourceIdentifierFromSubscriptionResponse'

Fields

eventSubscription :: Maybe EventSubscription

httpStatus :: Int

The response's http status code.

newRemoveSourceIdentifierFromSubscriptionResponse

Arguments

:: Int
$sel:httpStatus:RemoveSourceIdentifierFromSubscriptionResponse'

-> RemoveSourceIdentifierFromSubscriptionResponse

Create a value of RemoveSourceIdentifierFromSubscriptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSubscription:RemoveSourceIdentifierFromSubscriptionResponse', removeSourceIdentifierFromSubscriptionResponse_eventSubscription - Undocumented member.

$sel:httpStatus:RemoveSourceIdentifierFromSubscriptionResponse', removeSourceIdentifierFromSubscriptionResponse_httpStatus - The response's http status code.

Response Lenses

removeSourceIdentifierFromSubscriptionResponse_eventSubscription :: Lens' RemoveSourceIdentifierFromSubscriptionResponse (Maybe EventSubscription)

Undocumented member.

removeSourceIdentifierFromSubscriptionResponse_httpStatus :: Lens' RemoveSourceIdentifierFromSubscriptionResponse Int

The response's http status code.