Amazonka.Neptune.Types.DBClusterParameterGroup

Description

Documentation

data DBClusterParameterGroup

Contains the details of an Amazon Neptune DB cluster parameter group.

This data type is used as a response element in the DescribeDBClusterParameterGroups action.

See: newDBClusterParameterGroup smart constructor.

Constructors

DBClusterParameterGroup'

Fields

dbClusterParameterGroupArn :: Maybe Text

The Amazon Resource Name (ARN) for the DB cluster parameter group.

dbParameterGroupFamily :: Maybe Text

Provides the name of the DB parameter group family that this DB cluster parameter group is compatible with.

dbClusterParameterGroupName :: Maybe Text

Provides the name of the DB cluster parameter group.

description :: Maybe Text

Provides the customer-specified description for this DB cluster parameter group.

newDBClusterParameterGroup :: DBClusterParameterGroup

Create a value of DBClusterParameterGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterParameterGroupArn:DBClusterParameterGroup', dbClusterParameterGroup_dbClusterParameterGroupArn - The Amazon Resource Name (ARN) for the DB cluster parameter group.

$sel:dbParameterGroupFamily:DBClusterParameterGroup', dbClusterParameterGroup_dbParameterGroupFamily - Provides the name of the DB parameter group family that this DB cluster parameter group is compatible with.

$sel:dbClusterParameterGroupName:DBClusterParameterGroup', dbClusterParameterGroup_dbClusterParameterGroupName - Provides the name of the DB cluster parameter group.

$sel:description:DBClusterParameterGroup', dbClusterParameterGroup_description - Provides the customer-specified description for this DB cluster parameter group.

dbClusterParameterGroup_dbClusterParameterGroupArn :: Lens' DBClusterParameterGroup (Maybe Text)

The Amazon Resource Name (ARN) for the DB cluster parameter group.

dbClusterParameterGroup_dbParameterGroupFamily :: Lens' DBClusterParameterGroup (Maybe Text)

Provides the name of the DB parameter group family that this DB cluster parameter group is compatible with.

dbClusterParameterGroup_dbClusterParameterGroupName :: Lens' DBClusterParameterGroup (Maybe Text)

Provides the name of the DB cluster parameter group.

dbClusterParameterGroup_description :: Lens' DBClusterParameterGroup (Maybe Text)

Provides the customer-specified description for this DB cluster parameter group.