Amazonka.Neptune.Types.DBClusterSnapshot

Description

Documentation

data DBClusterSnapshot

Contains the details for an Amazon Neptune DB cluster snapshot

This data type is used as a response element in the DescribeDBClusterSnapshots action.

See: newDBClusterSnapshot smart constructor.

Constructors

DBClusterSnapshot'

Fields

engineVersion :: Maybe Text

Provides the version of the database engine for this DB cluster snapshot.

status :: Maybe Text

Specifies the status of this DB cluster snapshot.

storageEncrypted :: Maybe Bool

Specifies whether the DB cluster snapshot is encrypted.

dbClusterIdentifier :: Maybe Text

Specifies the DB cluster identifier of the DB cluster that this DB cluster snapshot was created from.

masterUsername :: Maybe Text

Not supported by Neptune.

iAMDatabaseAuthenticationEnabled :: Maybe Bool

True if mapping of Amazon Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

dbClusterSnapshotArn :: Maybe Text

The Amazon Resource Name (ARN) for the DB cluster snapshot.

vpcId :: Maybe Text

Provides the VPC ID associated with the DB cluster snapshot.

dbClusterSnapshotIdentifier :: Maybe Text

Specifies the identifier for a DB cluster snapshot. Must match the identifier of an existing snapshot.

After you restore a DB cluster using a DBClusterSnapshotIdentifier, you must specify the same DBClusterSnapshotIdentifier for any future updates to the DB cluster. When you specify this property for an update, the DB cluster is not restored from the snapshot again, and the data in the database is not changed.

However, if you don't specify the DBClusterSnapshotIdentifier, an empty DB cluster is created, and the original DB cluster is deleted. If you specify a property that is different from the previous snapshot restore property, the DB cluster is restored from the snapshot specified by the DBClusterSnapshotIdentifier, and the original DB cluster is deleted.

engine :: Maybe Text

Specifies the name of the database engine.

licenseModel :: Maybe Text

Provides the license model information for this DB cluster snapshot.

availabilityZones :: Maybe [Text]

Provides the list of EC2 Availability Zones that instances in the DB cluster snapshot can be restored in.

snapshotType :: Maybe Text

Provides the type of the DB cluster snapshot.

kmsKeyId :: Maybe Text

If StorageEncrypted is true, the Amazon KMS key identifier for the encrypted DB cluster snapshot.

snapshotCreateTime :: Maybe ISO8601

Provides the time when the snapshot was taken, in Universal Coordinated Time (UTC).

allocatedStorage :: Maybe Int

Specifies the allocated storage size in gibibytes (GiB).

sourceDBClusterSnapshotArn :: Maybe Text

If the DB cluster snapshot was copied from a source DB cluster snapshot, the Amazon Resource Name (ARN) for the source DB cluster snapshot, otherwise, a null value.

clusterCreateTime :: Maybe ISO8601

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

percentProgress :: Maybe Int

Specifies the percentage of the estimated data that has been transferred.

port :: Maybe Int

Specifies the port that the DB cluster was listening on at the time of the snapshot.

newDBClusterSnapshot :: DBClusterSnapshot

Create a value of DBClusterSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:DBClusterSnapshot', dbClusterSnapshot_engineVersion - Provides the version of the database engine for this DB cluster snapshot.

$sel:status:DBClusterSnapshot', dbClusterSnapshot_status - Specifies the status of this DB cluster snapshot.

$sel:storageEncrypted:DBClusterSnapshot', dbClusterSnapshot_storageEncrypted - Specifies whether the DB cluster snapshot is encrypted.

$sel:dbClusterIdentifier:DBClusterSnapshot', dbClusterSnapshot_dbClusterIdentifier - Specifies the DB cluster identifier of the DB cluster that this DB cluster snapshot was created from.

$sel:masterUsername:DBClusterSnapshot', dbClusterSnapshot_masterUsername - Not supported by Neptune.

$sel:iAMDatabaseAuthenticationEnabled:DBClusterSnapshot', dbClusterSnapshot_iAMDatabaseAuthenticationEnabled - True if mapping of Amazon Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

$sel:dbClusterSnapshotArn:DBClusterSnapshot', dbClusterSnapshot_dbClusterSnapshotArn - The Amazon Resource Name (ARN) for the DB cluster snapshot.

$sel:vpcId:DBClusterSnapshot', dbClusterSnapshot_vpcId - Provides the VPC ID associated with the DB cluster snapshot.

$sel:dbClusterSnapshotIdentifier:DBClusterSnapshot', dbClusterSnapshot_dbClusterSnapshotIdentifier - Specifies the identifier for a DB cluster snapshot. Must match the identifier of an existing snapshot.

After you restore a DB cluster using a DBClusterSnapshotIdentifier, you must specify the same DBClusterSnapshotIdentifier for any future updates to the DB cluster. When you specify this property for an update, the DB cluster is not restored from the snapshot again, and the data in the database is not changed.

However, if you don't specify the DBClusterSnapshotIdentifier, an empty DB cluster is created, and the original DB cluster is deleted. If you specify a property that is different from the previous snapshot restore property, the DB cluster is restored from the snapshot specified by the DBClusterSnapshotIdentifier, and the original DB cluster is deleted.

$sel:engine:DBClusterSnapshot', dbClusterSnapshot_engine - Specifies the name of the database engine.

$sel:licenseModel:DBClusterSnapshot', dbClusterSnapshot_licenseModel - Provides the license model information for this DB cluster snapshot.

$sel:availabilityZones:DBClusterSnapshot', dbClusterSnapshot_availabilityZones - Provides the list of EC2 Availability Zones that instances in the DB cluster snapshot can be restored in.

$sel:snapshotType:DBClusterSnapshot', dbClusterSnapshot_snapshotType - Provides the type of the DB cluster snapshot.

$sel:kmsKeyId:DBClusterSnapshot', dbClusterSnapshot_kmsKeyId - If StorageEncrypted is true, the Amazon KMS key identifier for the encrypted DB cluster snapshot.

$sel:snapshotCreateTime:DBClusterSnapshot', dbClusterSnapshot_snapshotCreateTime - Provides the time when the snapshot was taken, in Universal Coordinated Time (UTC).

$sel:allocatedStorage:DBClusterSnapshot', dbClusterSnapshot_allocatedStorage - Specifies the allocated storage size in gibibytes (GiB).

$sel:sourceDBClusterSnapshotArn:DBClusterSnapshot', dbClusterSnapshot_sourceDBClusterSnapshotArn - If the DB cluster snapshot was copied from a source DB cluster snapshot, the Amazon Resource Name (ARN) for the source DB cluster snapshot, otherwise, a null value.

$sel:clusterCreateTime:DBClusterSnapshot', dbClusterSnapshot_clusterCreateTime - Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

$sel:percentProgress:DBClusterSnapshot', dbClusterSnapshot_percentProgress - Specifies the percentage of the estimated data that has been transferred.

$sel:port:DBClusterSnapshot', dbClusterSnapshot_port - Specifies the port that the DB cluster was listening on at the time of the snapshot.

dbClusterSnapshot_engineVersion :: Lens' DBClusterSnapshot (Maybe Text)

Provides the version of the database engine for this DB cluster snapshot.

dbClusterSnapshot_status :: Lens' DBClusterSnapshot (Maybe Text)

Specifies the status of this DB cluster snapshot.

dbClusterSnapshot_storageEncrypted :: Lens' DBClusterSnapshot (Maybe Bool)

Specifies whether the DB cluster snapshot is encrypted.

dbClusterSnapshot_dbClusterIdentifier :: Lens' DBClusterSnapshot (Maybe Text)

Specifies the DB cluster identifier of the DB cluster that this DB cluster snapshot was created from.

dbClusterSnapshot_masterUsername :: Lens' DBClusterSnapshot (Maybe Text)

Not supported by Neptune.

dbClusterSnapshot_iAMDatabaseAuthenticationEnabled :: Lens' DBClusterSnapshot (Maybe Bool)

True if mapping of Amazon Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

dbClusterSnapshot_dbClusterSnapshotArn :: Lens' DBClusterSnapshot (Maybe Text)

The Amazon Resource Name (ARN) for the DB cluster snapshot.

dbClusterSnapshot_vpcId :: Lens' DBClusterSnapshot (Maybe Text)

Provides the VPC ID associated with the DB cluster snapshot.

dbClusterSnapshot_dbClusterSnapshotIdentifier :: Lens' DBClusterSnapshot (Maybe Text)

Specifies the identifier for a DB cluster snapshot. Must match the identifier of an existing snapshot.

After you restore a DB cluster using a DBClusterSnapshotIdentifier, you must specify the same DBClusterSnapshotIdentifier for any future updates to the DB cluster. When you specify this property for an update, the DB cluster is not restored from the snapshot again, and the data in the database is not changed.

However, if you don't specify the DBClusterSnapshotIdentifier, an empty DB cluster is created, and the original DB cluster is deleted. If you specify a property that is different from the previous snapshot restore property, the DB cluster is restored from the snapshot specified by the DBClusterSnapshotIdentifier, and the original DB cluster is deleted.

dbClusterSnapshot_engine :: Lens' DBClusterSnapshot (Maybe Text)

Specifies the name of the database engine.

dbClusterSnapshot_licenseModel :: Lens' DBClusterSnapshot (Maybe Text)

Provides the license model information for this DB cluster snapshot.

dbClusterSnapshot_availabilityZones :: Lens' DBClusterSnapshot (Maybe [Text])

Provides the list of EC2 Availability Zones that instances in the DB cluster snapshot can be restored in.

dbClusterSnapshot_snapshotType :: Lens' DBClusterSnapshot (Maybe Text)

Provides the type of the DB cluster snapshot.

dbClusterSnapshot_kmsKeyId :: Lens' DBClusterSnapshot (Maybe Text)

If StorageEncrypted is true, the Amazon KMS key identifier for the encrypted DB cluster snapshot.

dbClusterSnapshot_snapshotCreateTime :: Lens' DBClusterSnapshot (Maybe UTCTime)

Provides the time when the snapshot was taken, in Universal Coordinated Time (UTC).

dbClusterSnapshot_allocatedStorage :: Lens' DBClusterSnapshot (Maybe Int)

Specifies the allocated storage size in gibibytes (GiB).

dbClusterSnapshot_sourceDBClusterSnapshotArn :: Lens' DBClusterSnapshot (Maybe Text)

If the DB cluster snapshot was copied from a source DB cluster snapshot, the Amazon Resource Name (ARN) for the source DB cluster snapshot, otherwise, a null value.

dbClusterSnapshot_clusterCreateTime :: Lens' DBClusterSnapshot (Maybe UTCTime)

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

dbClusterSnapshot_percentProgress :: Lens' DBClusterSnapshot (Maybe Int)

Specifies the percentage of the estimated data that has been transferred.

dbClusterSnapshot_port :: Lens' DBClusterSnapshot (Maybe Int)

Specifies the port that the DB cluster was listening on at the time of the snapshot.