Amazonka.Neptune.Types.DBEngineVersion

Description

Documentation

data DBEngineVersion

This data type is used as a response element in the action DescribeDBEngineVersions.

See: newDBEngineVersion smart constructor.

Constructors

DBEngineVersion'

Fields

engineVersion :: Maybe Text

The version number of the database engine.

dbEngineVersionDescription :: Maybe Text

The description of the database engine version.

defaultCharacterSet :: Maybe CharacterSet

(Not supported by Neptune)

engine :: Maybe Text

The name of the database engine.

dbParameterGroupFamily :: Maybe Text

The name of the DB parameter group family for the database engine.

supportedCharacterSets :: Maybe [CharacterSet]

(Not supported by Neptune)

dbEngineDescription :: Maybe Text

The description of the database engine.

validUpgradeTarget :: Maybe [UpgradeTarget]

A list of engine versions that this database engine version can be upgraded to.

supportsLogExportsToCloudwatchLogs :: Maybe Bool

A value that indicates whether the engine version supports exporting the log types specified by ExportableLogTypes to CloudWatch Logs.

supportsReadReplica :: Maybe Bool

Indicates whether the database engine version supports read replicas.

supportedTimezones :: Maybe [Timezone]

A list of the time zones supported by this engine for the Timezone parameter of the CreateDBInstance action.

exportableLogTypes :: Maybe [Text]

The types of logs that the database engine has available for export to CloudWatch Logs.

newDBEngineVersion :: DBEngineVersion

Create a value of DBEngineVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:DBEngineVersion', dbEngineVersion_engineVersion - The version number of the database engine.

$sel:dbEngineVersionDescription:DBEngineVersion', dbEngineVersion_dbEngineVersionDescription - The description of the database engine version.

$sel:defaultCharacterSet:DBEngineVersion', dbEngineVersion_defaultCharacterSet - (Not supported by Neptune)

$sel:engine:DBEngineVersion', dbEngineVersion_engine - The name of the database engine.

$sel:dbParameterGroupFamily:DBEngineVersion', dbEngineVersion_dbParameterGroupFamily - The name of the DB parameter group family for the database engine.

$sel:supportedCharacterSets:DBEngineVersion', dbEngineVersion_supportedCharacterSets - (Not supported by Neptune)

$sel:dbEngineDescription:DBEngineVersion', dbEngineVersion_dbEngineDescription - The description of the database engine.

$sel:validUpgradeTarget:DBEngineVersion', dbEngineVersion_validUpgradeTarget - A list of engine versions that this database engine version can be upgraded to.

$sel:supportsLogExportsToCloudwatchLogs:DBEngineVersion', dbEngineVersion_supportsLogExportsToCloudwatchLogs - A value that indicates whether the engine version supports exporting the log types specified by ExportableLogTypes to CloudWatch Logs.

$sel:supportsReadReplica:DBEngineVersion', dbEngineVersion_supportsReadReplica - Indicates whether the database engine version supports read replicas.

$sel:supportedTimezones:DBEngineVersion', dbEngineVersion_supportedTimezones - A list of the time zones supported by this engine for the Timezone parameter of the CreateDBInstance action.

$sel:exportableLogTypes:DBEngineVersion', dbEngineVersion_exportableLogTypes - The types of logs that the database engine has available for export to CloudWatch Logs.

dbEngineVersion_engineVersion :: Lens' DBEngineVersion (Maybe Text)

The version number of the database engine.

dbEngineVersion_dbEngineVersionDescription :: Lens' DBEngineVersion (Maybe Text)

The description of the database engine version.

dbEngineVersion_defaultCharacterSet :: Lens' DBEngineVersion (Maybe CharacterSet)

(Not supported by Neptune)

dbEngineVersion_engine :: Lens' DBEngineVersion (Maybe Text)

The name of the database engine.

dbEngineVersion_dbParameterGroupFamily :: Lens' DBEngineVersion (Maybe Text)

The name of the DB parameter group family for the database engine.

dbEngineVersion_supportedCharacterSets :: Lens' DBEngineVersion (Maybe [CharacterSet])

(Not supported by Neptune)

dbEngineVersion_dbEngineDescription :: Lens' DBEngineVersion (Maybe Text)

The description of the database engine.

dbEngineVersion_validUpgradeTarget :: Lens' DBEngineVersion (Maybe [UpgradeTarget])

A list of engine versions that this database engine version can be upgraded to.

dbEngineVersion_supportsLogExportsToCloudwatchLogs :: Lens' DBEngineVersion (Maybe Bool)

A value that indicates whether the engine version supports exporting the log types specified by ExportableLogTypes to CloudWatch Logs.

dbEngineVersion_supportsReadReplica :: Lens' DBEngineVersion (Maybe Bool)

Indicates whether the database engine version supports read replicas.

dbEngineVersion_supportedTimezones :: Lens' DBEngineVersion (Maybe [Timezone])

A list of the time zones supported by this engine for the Timezone parameter of the CreateDBInstance action.

dbEngineVersion_exportableLogTypes :: Lens' DBEngineVersion (Maybe [Text])

The types of logs that the database engine has available for export to CloudWatch Logs.