Amazonka.NetworkFirewall.CreateFirewallPolicy

Description

Creates the firewall policy for the firewall according to the specifications.

An AWS Network Firewall firewall policy defines the behavior of a firewall, in a collection of stateless and stateful rule groups and other settings. You can use one firewall policy for multiple firewalls.

Creating a Request

data CreateFirewallPolicy

See: newCreateFirewallPolicy smart constructor.

Constructors

CreateFirewallPolicy'

Fields

description :: Maybe Text

A description of the firewall policy.

dryRun :: Maybe Bool

Indicates whether you want Network Firewall to just check the validity of the request, rather than run the request.

If set to TRUE, Network Firewall checks whether the request can run successfully, but doesn't actually make the requested changes. The call returns the value that the request would return if you ran it with dry run set to FALSE, but doesn't make additions or changes to your resources. This option allows you to make sure that you have the required permissions to run the request and that your request parameters are valid.

If set to FALSE, Network Firewall makes the requested changes to your resources.

tags :: Maybe (NonEmpty Tag)

The key:value pairs to associate with the resource.

firewallPolicyName :: Text

The descriptive name of the firewall policy. You can't change the name of a firewall policy after you create it.

firewallPolicy :: FirewallPolicy

The rule groups and policy actions to use in the firewall policy.

newCreateFirewallPolicy

Arguments

:: Text
$sel:firewallPolicyName:CreateFirewallPolicy'

-> FirewallPolicy
$sel:firewallPolicy:CreateFirewallPolicy'

-> CreateFirewallPolicy

Create a value of CreateFirewallPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateFirewallPolicy', createFirewallPolicy_description - A description of the firewall policy.

$sel:dryRun:CreateFirewallPolicy', createFirewallPolicy_dryRun - Indicates whether you want Network Firewall to just check the validity of the request, rather than run the request.

If set to TRUE, Network Firewall checks whether the request can run successfully, but doesn't actually make the requested changes. The call returns the value that the request would return if you ran it with dry run set to FALSE, but doesn't make additions or changes to your resources. This option allows you to make sure that you have the required permissions to run the request and that your request parameters are valid.

If set to FALSE, Network Firewall makes the requested changes to your resources.

$sel:tags:CreateFirewallPolicy', createFirewallPolicy_tags - The key:value pairs to associate with the resource.

$sel:firewallPolicyName:CreateFirewallPolicy', createFirewallPolicy_firewallPolicyName - The descriptive name of the firewall policy. You can't change the name of a firewall policy after you create it.

$sel:firewallPolicy:CreateFirewallPolicy', createFirewallPolicy_firewallPolicy - The rule groups and policy actions to use in the firewall policy.

Request Lenses

createFirewallPolicy_description :: Lens' CreateFirewallPolicy (Maybe Text)

A description of the firewall policy.

createFirewallPolicy_dryRun :: Lens' CreateFirewallPolicy (Maybe Bool)

Indicates whether you want Network Firewall to just check the validity of the request, rather than run the request.

If set to TRUE, Network Firewall checks whether the request can run successfully, but doesn't actually make the requested changes. The call returns the value that the request would return if you ran it with dry run set to FALSE, but doesn't make additions or changes to your resources. This option allows you to make sure that you have the required permissions to run the request and that your request parameters are valid.

If set to FALSE, Network Firewall makes the requested changes to your resources.

createFirewallPolicy_tags :: Lens' CreateFirewallPolicy (Maybe (NonEmpty Tag))

The key:value pairs to associate with the resource.

createFirewallPolicy_firewallPolicyName :: Lens' CreateFirewallPolicy Text

The descriptive name of the firewall policy. You can't change the name of a firewall policy after you create it.

createFirewallPolicy_firewallPolicy :: Lens' CreateFirewallPolicy FirewallPolicy

The rule groups and policy actions to use in the firewall policy.

Destructuring the Response

data CreateFirewallPolicyResponse

See: newCreateFirewallPolicyResponse smart constructor.

Constructors

CreateFirewallPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

updateToken :: Text

A token used for optimistic locking. Network Firewall returns a token to your requests that access the firewall policy. The token marks the state of the policy resource at the time of the request.

To make changes to the policy, you provide the token in your request. Network Firewall uses the token to ensure that the policy hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall policy again to get a current copy of it with current token. Reapply your changes as needed, then try the operation again using the new token.

firewallPolicyResponse :: FirewallPolicyResponse

The high-level properties of a firewall policy. This, along with the FirewallPolicy, define the policy. You can retrieve all objects for a firewall policy by calling DescribeFirewallPolicy.

newCreateFirewallPolicyResponse

Arguments

:: Int
$sel:httpStatus:CreateFirewallPolicyResponse'

-> Text
$sel:updateToken:CreateFirewallPolicyResponse'

-> FirewallPolicyResponse
$sel:firewallPolicyResponse:CreateFirewallPolicyResponse'

-> CreateFirewallPolicyResponse

Create a value of CreateFirewallPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateFirewallPolicyResponse', createFirewallPolicyResponse_httpStatus - The response's http status code.

$sel:updateToken:CreateFirewallPolicyResponse', createFirewallPolicyResponse_updateToken - A token used for optimistic locking. Network Firewall returns a token to your requests that access the firewall policy. The token marks the state of the policy resource at the time of the request.

To make changes to the policy, you provide the token in your request. Network Firewall uses the token to ensure that the policy hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall policy again to get a current copy of it with current token. Reapply your changes as needed, then try the operation again using the new token.

$sel:firewallPolicyResponse:CreateFirewallPolicyResponse', createFirewallPolicyResponse_firewallPolicyResponse - The high-level properties of a firewall policy. This, along with the FirewallPolicy, define the policy. You can retrieve all objects for a firewall policy by calling DescribeFirewallPolicy.

Response Lenses

createFirewallPolicyResponse_httpStatus :: Lens' CreateFirewallPolicyResponse Int

The response's http status code.

createFirewallPolicyResponse_updateToken :: Lens' CreateFirewallPolicyResponse Text

A token used for optimistic locking. Network Firewall returns a token to your requests that access the firewall policy. The token marks the state of the policy resource at the time of the request.

To make changes to the policy, you provide the token in your request. Network Firewall uses the token to ensure that the policy hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall policy again to get a current copy of it with current token. Reapply your changes as needed, then try the operation again using the new token.

createFirewallPolicyResponse_firewallPolicyResponse :: Lens' CreateFirewallPolicyResponse FirewallPolicyResponse

The high-level properties of a firewall policy. This, along with the FirewallPolicy, define the policy. You can retrieve all objects for a firewall policy by calling DescribeFirewallPolicy.