Amazonka.NetworkFirewall.DescribeRuleGroup

Description

Returns the data objects for the specified rule group.

Creating a Request

data DescribeRuleGroup

See: newDescribeRuleGroup smart constructor.

Constructors

DescribeRuleGroup'

Fields

ruleGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the rule group.

You must specify the ARN or the name, and you can specify both.

type' :: Maybe RuleGroupType

Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

This setting is required for requests that do not include the RuleGroupARN.

ruleGroupName :: Maybe Text

The descriptive name of the rule group. You can't change the name of a rule group after you create it.

You must specify the ARN or the name, and you can specify both.

newDescribeRuleGroup :: DescribeRuleGroup

Create a value of DescribeRuleGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleGroupArn:DescribeRuleGroup', describeRuleGroup_ruleGroupArn - The Amazon Resource Name (ARN) of the rule group.

You must specify the ARN or the name, and you can specify both.

$sel:type':DescribeRuleGroup', describeRuleGroup_type - Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

This setting is required for requests that do not include the RuleGroupARN.

$sel:ruleGroupName:DescribeRuleGroup', describeRuleGroup_ruleGroupName - The descriptive name of the rule group. You can't change the name of a rule group after you create it.

You must specify the ARN or the name, and you can specify both.

Request Lenses

describeRuleGroup_ruleGroupArn :: Lens' DescribeRuleGroup (Maybe Text)

The Amazon Resource Name (ARN) of the rule group.

You must specify the ARN or the name, and you can specify both.

describeRuleGroup_type :: Lens' DescribeRuleGroup (Maybe RuleGroupType)

Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

This setting is required for requests that do not include the RuleGroupARN.

describeRuleGroup_ruleGroupName :: Lens' DescribeRuleGroup (Maybe Text)

The descriptive name of the rule group. You can't change the name of a rule group after you create it.

You must specify the ARN or the name, and you can specify both.

Destructuring the Response

data DescribeRuleGroupResponse

See: newDescribeRuleGroupResponse smart constructor.

Constructors

DescribeRuleGroupResponse'

Fields

ruleGroup :: Maybe RuleGroup

The object that defines the rules in a rule group. This, along with RuleGroupResponse, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

AWS Network Firewall uses a rule group to inspect and control network traffic. You define stateless rule groups to inspect individual packets and you define stateful rule groups to inspect packets in the context of their traffic flow.

To use a rule group, you include it by reference in an Network Firewall firewall policy, then you use the policy in a firewall. You can reference a rule group from more than one firewall policy, and you can use a firewall policy in more than one firewall.

httpStatus :: Int

The response's http status code.

updateToken :: Text

A token used for optimistic locking. Network Firewall returns a token to your requests that access the rule group. The token marks the state of the rule group resource at the time of the request.

To make changes to the rule group, you provide the token in your request. Network Firewall uses the token to ensure that the rule group hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the rule group again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.

ruleGroupResponse :: RuleGroupResponse

The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

newDescribeRuleGroupResponse

Arguments

:: Int
$sel:httpStatus:DescribeRuleGroupResponse'

-> Text
$sel:updateToken:DescribeRuleGroupResponse'

-> RuleGroupResponse
$sel:ruleGroupResponse:DescribeRuleGroupResponse'

-> DescribeRuleGroupResponse

Create a value of DescribeRuleGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleGroup:DescribeRuleGroupResponse', describeRuleGroupResponse_ruleGroup - The object that defines the rules in a rule group. This, along with RuleGroupResponse, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

AWS Network Firewall uses a rule group to inspect and control network traffic. You define stateless rule groups to inspect individual packets and you define stateful rule groups to inspect packets in the context of their traffic flow.

To use a rule group, you include it by reference in an Network Firewall firewall policy, then you use the policy in a firewall. You can reference a rule group from more than one firewall policy, and you can use a firewall policy in more than one firewall.

$sel:httpStatus:DescribeRuleGroupResponse', describeRuleGroupResponse_httpStatus - The response's http status code.

$sel:updateToken:DescribeRuleGroupResponse', describeRuleGroupResponse_updateToken - A token used for optimistic locking. Network Firewall returns a token to your requests that access the rule group. The token marks the state of the rule group resource at the time of the request.

To make changes to the rule group, you provide the token in your request. Network Firewall uses the token to ensure that the rule group hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the rule group again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.

$sel:ruleGroupResponse:DescribeRuleGroupResponse', describeRuleGroupResponse_ruleGroupResponse - The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

Response Lenses

describeRuleGroupResponse_ruleGroup :: Lens' DescribeRuleGroupResponse (Maybe RuleGroup)

The object that defines the rules in a rule group. This, along with RuleGroupResponse, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

AWS Network Firewall uses a rule group to inspect and control network traffic. You define stateless rule groups to inspect individual packets and you define stateful rule groups to inspect packets in the context of their traffic flow.

To use a rule group, you include it by reference in an Network Firewall firewall policy, then you use the policy in a firewall. You can reference a rule group from more than one firewall policy, and you can use a firewall policy in more than one firewall.

describeRuleGroupResponse_httpStatus :: Lens' DescribeRuleGroupResponse Int

The response's http status code.

describeRuleGroupResponse_updateToken :: Lens' DescribeRuleGroupResponse Text

A token used for optimistic locking. Network Firewall returns a token to your requests that access the rule group. The token marks the state of the rule group resource at the time of the request.

To make changes to the rule group, you provide the token in your request. Network Firewall uses the token to ensure that the rule group hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the rule group again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.

describeRuleGroupResponse_ruleGroupResponse :: Lens' DescribeRuleGroupResponse RuleGroupResponse

The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.