Amazonka.NetworkFirewall.Types.FirewallMetadata

Description

Documentation

data FirewallMetadata

High-level information about a firewall, returned by operations like create and describe. You can use the information provided in the metadata to retrieve and manage a firewall.

See: newFirewallMetadata smart constructor.

Constructors

FirewallMetadata'

Fields

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

newFirewallMetadata :: FirewallMetadata

Create a value of FirewallMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallArn:FirewallMetadata', firewallMetadata_firewallArn - The Amazon Resource Name (ARN) of the firewall.

$sel:firewallName:FirewallMetadata', firewallMetadata_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

firewallMetadata_firewallArn :: Lens' FirewallMetadata (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

firewallMetadata_firewallName :: Lens' FirewallMetadata (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.