Amazonka.NetworkFirewall.Types.FirewallPolicyResponse

Description

Documentation

data FirewallPolicyResponse

The high-level properties of a firewall policy. This, along with the FirewallPolicy, define the policy. You can retrieve all objects for a firewall policy by calling DescribeFirewallPolicy.

See: newFirewallPolicyResponse smart constructor.

Constructors

FirewallPolicyResponse'

Fields

consumedStatelessRuleCapacity :: Maybe Int

The number of capacity units currently consumed by the policy's stateless rules.

numberOfAssociations :: Maybe Int

The number of firewalls that are associated with this firewall policy.

firewallPolicyStatus :: Maybe ResourceStatus

The current status of the firewall policy. You can retrieve this for a firewall policy by calling DescribeFirewallPolicy and providing the firewall policy's name or ARN.

consumedStatefulRuleCapacity :: Maybe Int

The number of capacity units currently consumed by the policy's stateful rules.

description :: Maybe Text

A description of the firewall policy.

tags :: Maybe (NonEmpty Tag)

The key:value pairs to associate with the resource.

firewallPolicyName :: Text

The descriptive name of the firewall policy. You can't change the name of a firewall policy after you create it.

firewallPolicyArn :: Text

The Amazon Resource Name (ARN) of the firewall policy.

If this response is for a create request that had DryRun set to TRUE, then this ARN is a placeholder that isn't attached to a valid resource.

firewallPolicyId :: Text

The unique identifier for the firewall policy.

newFirewallPolicyResponse

Arguments

:: Text
$sel:firewallPolicyName:FirewallPolicyResponse'

-> Text
$sel:firewallPolicyArn:FirewallPolicyResponse'

-> Text
$sel:firewallPolicyId:FirewallPolicyResponse'

-> FirewallPolicyResponse

Create a value of FirewallPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:consumedStatelessRuleCapacity:FirewallPolicyResponse', firewallPolicyResponse_consumedStatelessRuleCapacity - The number of capacity units currently consumed by the policy's stateless rules.

$sel:numberOfAssociations:FirewallPolicyResponse', firewallPolicyResponse_numberOfAssociations - The number of firewalls that are associated with this firewall policy.

$sel:firewallPolicyStatus:FirewallPolicyResponse', firewallPolicyResponse_firewallPolicyStatus - The current status of the firewall policy. You can retrieve this for a firewall policy by calling DescribeFirewallPolicy and providing the firewall policy's name or ARN.

$sel:consumedStatefulRuleCapacity:FirewallPolicyResponse', firewallPolicyResponse_consumedStatefulRuleCapacity - The number of capacity units currently consumed by the policy's stateful rules.

$sel:description:FirewallPolicyResponse', firewallPolicyResponse_description - A description of the firewall policy.

$sel:tags:FirewallPolicyResponse', firewallPolicyResponse_tags - The key:value pairs to associate with the resource.

$sel:firewallPolicyName:FirewallPolicyResponse', firewallPolicyResponse_firewallPolicyName - The descriptive name of the firewall policy. You can't change the name of a firewall policy after you create it.

$sel:firewallPolicyArn:FirewallPolicyResponse', firewallPolicyResponse_firewallPolicyArn - The Amazon Resource Name (ARN) of the firewall policy.

If this response is for a create request that had DryRun set to TRUE, then this ARN is a placeholder that isn't attached to a valid resource.

$sel:firewallPolicyId:FirewallPolicyResponse', firewallPolicyResponse_firewallPolicyId - The unique identifier for the firewall policy.

firewallPolicyResponse_consumedStatelessRuleCapacity :: Lens' FirewallPolicyResponse (Maybe Int)

The number of capacity units currently consumed by the policy's stateless rules.

firewallPolicyResponse_numberOfAssociations :: Lens' FirewallPolicyResponse (Maybe Int)

The number of firewalls that are associated with this firewall policy.

firewallPolicyResponse_firewallPolicyStatus :: Lens' FirewallPolicyResponse (Maybe ResourceStatus)

The current status of the firewall policy. You can retrieve this for a firewall policy by calling DescribeFirewallPolicy and providing the firewall policy's name or ARN.

firewallPolicyResponse_consumedStatefulRuleCapacity :: Lens' FirewallPolicyResponse (Maybe Int)

The number of capacity units currently consumed by the policy's stateful rules.

firewallPolicyResponse_description :: Lens' FirewallPolicyResponse (Maybe Text)

A description of the firewall policy.

firewallPolicyResponse_tags :: Lens' FirewallPolicyResponse (Maybe (NonEmpty Tag))

The key:value pairs to associate with the resource.

firewallPolicyResponse_firewallPolicyName :: Lens' FirewallPolicyResponse Text

The descriptive name of the firewall policy. You can't change the name of a firewall policy after you create it.

firewallPolicyResponse_firewallPolicyArn :: Lens' FirewallPolicyResponse Text

The Amazon Resource Name (ARN) of the firewall policy.

If this response is for a create request that had DryRun set to TRUE, then this ARN is a placeholder that isn't attached to a valid resource.

firewallPolicyResponse_firewallPolicyId :: Lens' FirewallPolicyResponse Text

The unique identifier for the firewall policy.