Amazonka.NetworkFirewall.Types.Header

Description

Documentation

data Header

The basic rule criteria for AWS Network Firewall to use to inspect packet headers in stateful traffic flow inspection. Traffic flows that match the criteria are a match for the corresponding StatefulRule.

See: newHeader smart constructor.

Constructors

Header'

Fields

protocol :: StatefulRuleProtocol

The protocol to inspect for. To specify all, you can use IP, because all traffic on AWS and on the internet is IP.

source :: Text

The source IP address or address range to inspect for, in CIDR notation. To match with any address, specify ANY.

Specify an IP address or a block of IP addresses in Classless Inter-Domain Routing (CIDR) notation. Network Firewall supports all address ranges for IPv4.

Examples:

To configure Network Firewall to inspect for the IP address 192.0.2.44, specify 192.0.2.44/32.

To configure Network Firewall to inspect for IP addresses from 192.0.2.0 to 192.0.2.255, specify 192.0.2.0/24.

For more information about CIDR notation, see the Wikipedia entry Classless Inter-Domain Routing.

sourcePort :: Text

The source port to inspect for. You can specify an individual port, for example 1994 and you can specify a port range, for example 1990:1994. To match with any port, specify ANY.

direction :: StatefulRuleDirection

The direction of traffic flow to inspect. If set to ANY, the inspection matches bidirectional traffic, both from the source to the destination and from the destination to the source. If set to FORWARD, the inspection only matches traffic going from the source to the destination.

destination :: Text

The destination IP address or address range to inspect for, in CIDR notation. To match with any address, specify ANY.

Specify an IP address or a block of IP addresses in Classless Inter-Domain Routing (CIDR) notation. Network Firewall supports all address ranges for IPv4.

Examples:

To configure Network Firewall to inspect for the IP address 192.0.2.44, specify 192.0.2.44/32.

To configure Network Firewall to inspect for IP addresses from 192.0.2.0 to 192.0.2.255, specify 192.0.2.0/24.

For more information about CIDR notation, see the Wikipedia entry Classless Inter-Domain Routing.

destinationPort :: Text

The destination port to inspect for. You can specify an individual port, for example 1994 and you can specify a port range, for example 1990:1994. To match with any port, specify ANY.

newHeader

Arguments

:: StatefulRuleProtocol
$sel:protocol:Header'

-> Text
$sel:source:Header'

-> Text
$sel:sourcePort:Header'

-> StatefulRuleDirection
$sel:direction:Header'

-> Text
$sel:destination:Header'

-> Text
$sel:destinationPort:Header'

-> Header

Create a value of Header with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocol:Header', header_protocol - The protocol to inspect for. To specify all, you can use IP, because all traffic on AWS and on the internet is IP.

$sel:source:Header', header_source - The source IP address or address range to inspect for, in CIDR notation. To match with any address, specify ANY.

Specify an IP address or a block of IP addresses in Classless Inter-Domain Routing (CIDR) notation. Network Firewall supports all address ranges for IPv4.

Examples:

To configure Network Firewall to inspect for the IP address 192.0.2.44, specify 192.0.2.44/32.

To configure Network Firewall to inspect for IP addresses from 192.0.2.0 to 192.0.2.255, specify 192.0.2.0/24.

For more information about CIDR notation, see the Wikipedia entry Classless Inter-Domain Routing.

$sel:sourcePort:Header', header_sourcePort - The source port to inspect for. You can specify an individual port, for example 1994 and you can specify a port range, for example 1990:1994. To match with any port, specify ANY.

$sel:direction:Header', header_direction - The direction of traffic flow to inspect. If set to ANY, the inspection matches bidirectional traffic, both from the source to the destination and from the destination to the source. If set to FORWARD, the inspection only matches traffic going from the source to the destination.

$sel:destination:Header', header_destination - The destination IP address or address range to inspect for, in CIDR notation. To match with any address, specify ANY.

Specify an IP address or a block of IP addresses in Classless Inter-Domain Routing (CIDR) notation. Network Firewall supports all address ranges for IPv4.

Examples:

To configure Network Firewall to inspect for the IP address 192.0.2.44, specify 192.0.2.44/32.

To configure Network Firewall to inspect for IP addresses from 192.0.2.0 to 192.0.2.255, specify 192.0.2.0/24.

For more information about CIDR notation, see the Wikipedia entry Classless Inter-Domain Routing.

$sel:destinationPort:Header', header_destinationPort - The destination port to inspect for. You can specify an individual port, for example 1994 and you can specify a port range, for example 1990:1994. To match with any port, specify ANY.

header_protocol :: Lens' Header StatefulRuleProtocol

The protocol to inspect for. To specify all, you can use IP, because all traffic on AWS and on the internet is IP.

header_source :: Lens' Header Text

The source IP address or address range to inspect for, in CIDR notation. To match with any address, specify ANY.

Specify an IP address or a block of IP addresses in Classless Inter-Domain Routing (CIDR) notation. Network Firewall supports all address ranges for IPv4.

Examples:

To configure Network Firewall to inspect for the IP address 192.0.2.44, specify 192.0.2.44/32.

To configure Network Firewall to inspect for IP addresses from 192.0.2.0 to 192.0.2.255, specify 192.0.2.0/24.

For more information about CIDR notation, see the Wikipedia entry Classless Inter-Domain Routing.

header_sourcePort :: Lens' Header Text

The source port to inspect for. You can specify an individual port, for example 1994 and you can specify a port range, for example 1990:1994. To match with any port, specify ANY.

header_direction :: Lens' Header StatefulRuleDirection

The direction of traffic flow to inspect. If set to ANY, the inspection matches bidirectional traffic, both from the source to the destination and from the destination to the source. If set to FORWARD, the inspection only matches traffic going from the source to the destination.

header_destination :: Lens' Header Text

The destination IP address or address range to inspect for, in CIDR notation. To match with any address, specify ANY.

Specify an IP address or a block of IP addresses in Classless Inter-Domain Routing (CIDR) notation. Network Firewall supports all address ranges for IPv4.

Examples:

To configure Network Firewall to inspect for the IP address 192.0.2.44, specify 192.0.2.44/32.

To configure Network Firewall to inspect for IP addresses from 192.0.2.0 to 192.0.2.255, specify 192.0.2.0/24.

For more information about CIDR notation, see the Wikipedia entry Classless Inter-Domain Routing.

header_destinationPort :: Lens' Header Text

The destination port to inspect for. You can specify an individual port, for example 1994 and you can specify a port range, for example 1990:1994. To match with any port, specify ANY.