Amazonka.NetworkFirewall.Types.MatchAttributes

Description

Documentation

data MatchAttributes

Criteria for Network Firewall to use to inspect an individual packet in stateless rule inspection. Each match attributes set can include one or more items such as IP address, CIDR range, port number, protocol, and TCP flags.

See: newMatchAttributes smart constructor.

Constructors

MatchAttributes'

Fields

protocols :: Maybe [Natural]

The protocols to inspect for, specified using each protocol's assigned internet protocol number (IANA). If not specified, this matches with any protocol.

tCPFlags :: Maybe [TCPFlagField]

The TCP flags and masks to inspect for. If not specified, this matches with any settings. This setting is only used for protocol 6 (TCP).

destinationPorts :: Maybe [PortRange]

The destination ports to inspect for. If not specified, this matches with any destination port. This setting is only used for protocols 6 (TCP) and 17 (UDP).

You can specify individual ports, for example 1994 and you can specify port ranges, for example 1990:1994.

sources :: Maybe [Address]

The source IP addresses and address ranges to inspect for, in CIDR notation. If not specified, this matches with any source address.

sourcePorts :: Maybe [PortRange]

The source ports to inspect for. If not specified, this matches with any source port. This setting is only used for protocols 6 (TCP) and 17 (UDP).

You can specify individual ports, for example 1994 and you can specify port ranges, for example 1990:1994.

destinations :: Maybe [Address]

The destination IP addresses and address ranges to inspect for, in CIDR notation. If not specified, this matches with any destination address.

newMatchAttributes :: MatchAttributes

Create a value of MatchAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocols:MatchAttributes', matchAttributes_protocols - The protocols to inspect for, specified using each protocol's assigned internet protocol number (IANA). If not specified, this matches with any protocol.

$sel:tCPFlags:MatchAttributes', matchAttributes_tCPFlags - The TCP flags and masks to inspect for. If not specified, this matches with any settings. This setting is only used for protocol 6 (TCP).

$sel:destinationPorts:MatchAttributes', matchAttributes_destinationPorts - The destination ports to inspect for. If not specified, this matches with any destination port. This setting is only used for protocols 6 (TCP) and 17 (UDP).

You can specify individual ports, for example 1994 and you can specify port ranges, for example 1990:1994.

$sel:sources:MatchAttributes', matchAttributes_sources - The source IP addresses and address ranges to inspect for, in CIDR notation. If not specified, this matches with any source address.

$sel:sourcePorts:MatchAttributes', matchAttributes_sourcePorts - The source ports to inspect for. If not specified, this matches with any source port. This setting is only used for protocols 6 (TCP) and 17 (UDP).

You can specify individual ports, for example 1994 and you can specify port ranges, for example 1990:1994.

$sel:destinations:MatchAttributes', matchAttributes_destinations - The destination IP addresses and address ranges to inspect for, in CIDR notation. If not specified, this matches with any destination address.

matchAttributes_protocols :: Lens' MatchAttributes (Maybe [Natural])

The protocols to inspect for, specified using each protocol's assigned internet protocol number (IANA). If not specified, this matches with any protocol.

matchAttributes_tCPFlags :: Lens' MatchAttributes (Maybe [TCPFlagField])

The TCP flags and masks to inspect for. If not specified, this matches with any settings. This setting is only used for protocol 6 (TCP).

matchAttributes_destinationPorts :: Lens' MatchAttributes (Maybe [PortRange])

The destination ports to inspect for. If not specified, this matches with any destination port. This setting is only used for protocols 6 (TCP) and 17 (UDP).

You can specify individual ports, for example 1994 and you can specify port ranges, for example 1990:1994.

matchAttributes_sources :: Lens' MatchAttributes (Maybe [Address])

The source IP addresses and address ranges to inspect for, in CIDR notation. If not specified, this matches with any source address.

matchAttributes_sourcePorts :: Lens' MatchAttributes (Maybe [PortRange])

The source ports to inspect for. If not specified, this matches with any source port. This setting is only used for protocols 6 (TCP) and 17 (UDP).

You can specify individual ports, for example 1994 and you can specify port ranges, for example 1990:1994.

matchAttributes_destinations :: Lens' MatchAttributes (Maybe [Address])

The destination IP addresses and address ranges to inspect for, in CIDR notation. If not specified, this matches with any destination address.