Amazonka.NetworkFirewall.Types.PerObjectStatus

Description

Documentation

data PerObjectStatus

Provides configuration status for a single policy or rule group that is used for a firewall endpoint. Network Firewall provides each endpoint with the rules that are configured in the firewall policy. Each time you add a subnet or modify the associated firewall policy, Network Firewall synchronizes the rules in the endpoint, so it can properly filter network traffic. This is part of a SyncState for a firewall.

See: newPerObjectStatus smart constructor.

Constructors

PerObjectStatus'

Fields

updateToken :: Maybe Text

The current version of the object that is either in sync or pending synchronization.

syncStatus :: Maybe PerObjectSyncStatus

Indicates whether this object is in sync with the version indicated in the update token.

newPerObjectStatus :: PerObjectStatus

Create a value of PerObjectStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateToken:PerObjectStatus', perObjectStatus_updateToken - The current version of the object that is either in sync or pending synchronization.

$sel:syncStatus:PerObjectStatus', perObjectStatus_syncStatus - Indicates whether this object is in sync with the version indicated in the update token.

perObjectStatus_updateToken :: Lens' PerObjectStatus (Maybe Text)

The current version of the object that is either in sync or pending synchronization.

perObjectStatus_syncStatus :: Lens' PerObjectStatus (Maybe PerObjectSyncStatus)

Indicates whether this object is in sync with the version indicated in the update token.