Amazonka.NetworkFirewall.Types.RuleVariables

Description

Documentation

data RuleVariables

Settings that are available for use in the rules in the RuleGroup where this is defined.

See: newRuleVariables smart constructor.

Constructors

RuleVariables'

Fields

portSets :: Maybe (HashMap Text PortSet)

A list of port ranges.

iPSets :: Maybe (HashMap Text IPSet)

A list of IP addresses and address ranges, in CIDR notation.

newRuleVariables :: RuleVariables

Create a value of RuleVariables with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portSets:RuleVariables', ruleVariables_portSets - A list of port ranges.

$sel:iPSets:RuleVariables', ruleVariables_iPSets - A list of IP addresses and address ranges, in CIDR notation.

ruleVariables_portSets :: Lens' RuleVariables (Maybe (HashMap Text PortSet))

A list of port ranges.

ruleVariables_iPSets :: Lens' RuleVariables (Maybe (HashMap Text IPSet))

A list of IP addresses and address ranges, in CIDR notation.