Amazonka.NetworkFirewall.Types.RulesSourceList

Description

Documentation

data RulesSourceList

Stateful inspection criteria for a domain list rule group.

For HTTPS traffic, domain filtering is SNI-based. It uses the server name indicator extension of the TLS handshake.

By default, Network Firewall domain list inspection only includes traffic coming from the VPC where you deploy the firewall. To inspect traffic from IP addresses outside of the deployment VPC, you set the HOME_NET rule variable to include the CIDR range of the deployment VPC plus the other CIDR ranges. For more information, see RuleVariables in this guide and Stateful domain list rule groups in AWS Network Firewall in the Network Firewall Developer Guide.

See: newRulesSourceList smart constructor.

Constructors

RulesSourceList'

Fields

targets :: [Text]

The domains that you want to inspect for in your traffic flows. To provide multiple domains, separate them with commas. Valid domain specifications are the following:

Explicit names. For example, abc.example.com matches only the domain abc.example.com.

Names that use a domain wildcard, which you indicate with an initial '.'. For example,.example.com matches example.com and matches all subdomains of example.com, such as abc.example.com and www.example.com.

targetTypes :: [TargetType]

The protocols you want to inspect. Specify TLS_SNI for HTTPS. Specify HTTP_HOST for HTTP. You can specify either or both.

generatedRulesType :: GeneratedRulesType

Whether you want to allow or deny access to the domains in your target list.

newRulesSourceList

Arguments

:: GeneratedRulesType
$sel:generatedRulesType:RulesSourceList'

-> RulesSourceList

Create a value of RulesSourceList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targets:RulesSourceList', rulesSourceList_targets - The domains that you want to inspect for in your traffic flows. To provide multiple domains, separate them with commas. Valid domain specifications are the following:

Explicit names. For example, abc.example.com matches only the domain abc.example.com.

Names that use a domain wildcard, which you indicate with an initial '.'. For example,.example.com matches example.com and matches all subdomains of example.com, such as abc.example.com and www.example.com.

$sel:targetTypes:RulesSourceList', rulesSourceList_targetTypes - The protocols you want to inspect. Specify TLS_SNI for HTTPS. Specify HTTP_HOST for HTTP. You can specify either or both.

$sel:generatedRulesType:RulesSourceList', rulesSourceList_generatedRulesType - Whether you want to allow or deny access to the domains in your target list.

rulesSourceList_targets :: Lens' RulesSourceList [Text]

The domains that you want to inspect for in your traffic flows. To provide multiple domains, separate them with commas. Valid domain specifications are the following:

Explicit names. For example, abc.example.com matches only the domain abc.example.com.

Names that use a domain wildcard, which you indicate with an initial '.'. For example,.example.com matches example.com and matches all subdomains of example.com, such as abc.example.com and www.example.com.

rulesSourceList_targetTypes :: Lens' RulesSourceList [TargetType]

The protocols you want to inspect. Specify TLS_SNI for HTTPS. Specify HTTP_HOST for HTTP. You can specify either or both.

rulesSourceList_generatedRulesType :: Lens' RulesSourceList GeneratedRulesType

Whether you want to allow or deny access to the domains in your target list.