Amazonka.NetworkFirewall.Types.StatefulRuleOptions

Description

Documentation

data StatefulRuleOptions

Additional options governing how Network Firewall handles the rule group. You can only use these for stateful rule groups.

See: newStatefulRuleOptions smart constructor.

Constructors

StatefulRuleOptions'

Fields

ruleOrder :: Maybe RuleOrder

Indicates how to manage the order of the rule evaluation for the rule group. By default, Network Firewall leaves the rule evaluation order up to the Suricata rule processing engine. If you set this to STRICT_ORDER, your rules are evaluated in the exact order that they're listed in your Suricata rules string.

newStatefulRuleOptions :: StatefulRuleOptions

Create a value of StatefulRuleOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleOrder:StatefulRuleOptions', statefulRuleOptions_ruleOrder - Indicates how to manage the order of the rule evaluation for the rule group. By default, Network Firewall leaves the rule evaluation order up to the Suricata rule processing engine. If you set this to STRICT_ORDER, your rules are evaluated in the exact order that they're listed in your Suricata rules string.

statefulRuleOptions_ruleOrder :: Lens' StatefulRuleOptions (Maybe RuleOrder)

Indicates how to manage the order of the rule evaluation for the rule group. By default, Network Firewall leaves the rule evaluation order up to the Suricata rule processing engine. If you set this to STRICT_ORDER, your rules are evaluated in the exact order that they're listed in your Suricata rules string.