Amazonka.NetworkFirewall.Types.StatelessRule

Description

Documentation

data StatelessRule

A single stateless rule. This is used in StatelessRulesAndCustomActions.

See: newStatelessRule smart constructor.

Constructors

StatelessRule'

Fields

ruleDefinition :: RuleDefinition

Defines the stateless 5-tuple packet inspection criteria and the action to take on a packet that matches the criteria.

priority :: Natural

Indicates the order in which to run this rule relative to all of the rules that are defined for a stateless rule group. Network Firewall evaluates the rules in a rule group starting with the lowest priority setting. You must ensure that the priority settings are unique for the rule group.

Each stateless rule group uses exactly one StatelessRulesAndCustomActions object, and each StatelessRulesAndCustomActions contains exactly one StatelessRules object. To ensure unique priority settings for your rule groups, set unique priorities for the stateless rules that you define inside any single StatelessRules object.

You can change the priority settings of your rules at any time. To make it easier to insert rules later, number them so there's a wide range in between, for example use 100, 200, and so on.

newStatelessRule

Arguments

:: RuleDefinition
$sel:ruleDefinition:StatelessRule'

-> Natural
$sel:priority:StatelessRule'

-> StatelessRule

Create a value of StatelessRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleDefinition:StatelessRule', statelessRule_ruleDefinition - Defines the stateless 5-tuple packet inspection criteria and the action to take on a packet that matches the criteria.

$sel:priority:StatelessRule', statelessRule_priority - Indicates the order in which to run this rule relative to all of the rules that are defined for a stateless rule group. Network Firewall evaluates the rules in a rule group starting with the lowest priority setting. You must ensure that the priority settings are unique for the rule group.

Each stateless rule group uses exactly one StatelessRulesAndCustomActions object, and each StatelessRulesAndCustomActions contains exactly one StatelessRules object. To ensure unique priority settings for your rule groups, set unique priorities for the stateless rules that you define inside any single StatelessRules object.

You can change the priority settings of your rules at any time. To make it easier to insert rules later, number them so there's a wide range in between, for example use 100, 200, and so on.

statelessRule_ruleDefinition :: Lens' StatelessRule RuleDefinition

Defines the stateless 5-tuple packet inspection criteria and the action to take on a packet that matches the criteria.

statelessRule_priority :: Lens' StatelessRule Natural

Indicates the order in which to run this rule relative to all of the rules that are defined for a stateless rule group. Network Firewall evaluates the rules in a rule group starting with the lowest priority setting. You must ensure that the priority settings are unique for the rule group.

Each stateless rule group uses exactly one StatelessRulesAndCustomActions object, and each StatelessRulesAndCustomActions contains exactly one StatelessRules object. To ensure unique priority settings for your rule groups, set unique priorities for the stateless rules that you define inside any single StatelessRules object.

You can change the priority settings of your rules at any time. To make it easier to insert rules later, number them so there's a wide range in between, for example use 100, 200, and so on.