Amazonka.NetworkFirewall.UpdateFirewallDeleteProtection

Description

Modifies the flag, DeleteProtection, which indicates whether it is possible to delete the firewall. If the flag is set to TRUE, the firewall is protected against deletion. This setting helps protect against accidentally deleting a firewall that's in use.

Creating a Request

data UpdateFirewallDeleteProtection

See: newUpdateFirewallDeleteProtection smart constructor.

Constructors

UpdateFirewallDeleteProtection'

Fields

updateToken :: Maybe Text

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

deleteProtection :: Bool

A flag indicating whether it is possible to delete the firewall. A setting of TRUE indicates that the firewall is protected against deletion. Use this setting to protect against accidentally deleting a firewall that is in use. When you create a firewall, the operation initializes this flag to TRUE.

newUpdateFirewallDeleteProtection

Arguments

:: Bool
$sel:deleteProtection:UpdateFirewallDeleteProtection'

-> UpdateFirewallDeleteProtection

Create a value of UpdateFirewallDeleteProtection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateToken:UpdateFirewallDeleteProtection', updateFirewallDeleteProtection_updateToken - An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

$sel:firewallArn:UpdateFirewallDeleteProtection', updateFirewallDeleteProtection_firewallArn - The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

$sel:firewallName:UpdateFirewallDeleteProtection', updateFirewallDeleteProtection_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

$sel:deleteProtection:UpdateFirewallDeleteProtection', updateFirewallDeleteProtection_deleteProtection - A flag indicating whether it is possible to delete the firewall. A setting of TRUE indicates that the firewall is protected against deletion. Use this setting to protect against accidentally deleting a firewall that is in use. When you create a firewall, the operation initializes this flag to TRUE.

Request Lenses

updateFirewallDeleteProtection_updateToken :: Lens' UpdateFirewallDeleteProtection (Maybe Text)

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

updateFirewallDeleteProtection_firewallArn :: Lens' UpdateFirewallDeleteProtection (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

updateFirewallDeleteProtection_firewallName :: Lens' UpdateFirewallDeleteProtection (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

updateFirewallDeleteProtection_deleteProtection :: Lens' UpdateFirewallDeleteProtection Bool

A flag indicating whether it is possible to delete the firewall. A setting of TRUE indicates that the firewall is protected against deletion. Use this setting to protect against accidentally deleting a firewall that is in use. When you create a firewall, the operation initializes this flag to TRUE.

Destructuring the Response

data UpdateFirewallDeleteProtectionResponse

See: newUpdateFirewallDeleteProtectionResponse smart constructor.

Constructors

UpdateFirewallDeleteProtectionResponse'

Fields

updateToken :: Maybe Text

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

deleteProtection :: Maybe Bool

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

httpStatus :: Int

The response's http status code.

newUpdateFirewallDeleteProtectionResponse

Arguments

:: Int
$sel:httpStatus:UpdateFirewallDeleteProtectionResponse'

-> UpdateFirewallDeleteProtectionResponse

Create a value of UpdateFirewallDeleteProtectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateToken:UpdateFirewallDeleteProtection', updateFirewallDeleteProtectionResponse_updateToken - An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

$sel:firewallArn:UpdateFirewallDeleteProtection', updateFirewallDeleteProtectionResponse_firewallArn - The Amazon Resource Name (ARN) of the firewall.

$sel:deleteProtection:UpdateFirewallDeleteProtection', updateFirewallDeleteProtectionResponse_deleteProtection -

$sel:firewallName:UpdateFirewallDeleteProtection', updateFirewallDeleteProtectionResponse_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

$sel:httpStatus:UpdateFirewallDeleteProtectionResponse', updateFirewallDeleteProtectionResponse_httpStatus - The response's http status code.

Response Lenses

updateFirewallDeleteProtectionResponse_updateToken :: Lens' UpdateFirewallDeleteProtectionResponse (Maybe Text)

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

updateFirewallDeleteProtectionResponse_firewallArn :: Lens' UpdateFirewallDeleteProtectionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

updateFirewallDeleteProtectionResponse_deleteProtection :: Lens' UpdateFirewallDeleteProtectionResponse (Maybe Bool)

updateFirewallDeleteProtectionResponse_firewallName :: Lens' UpdateFirewallDeleteProtectionResponse (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

updateFirewallDeleteProtectionResponse_httpStatus :: Lens' UpdateFirewallDeleteProtectionResponse Int

The response's http status code.