Amazonka.NetworkFirewall.UpdateFirewallPolicyChangeProtection

Description

Creating a Request

data UpdateFirewallPolicyChangeProtection

See: newUpdateFirewallPolicyChangeProtection smart constructor.

Constructors

UpdateFirewallPolicyChangeProtection'

Fields

updateToken :: Maybe Text

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

firewallPolicyChangeProtection :: Bool

A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

newUpdateFirewallPolicyChangeProtection

Arguments

:: Bool
$sel:firewallPolicyChangeProtection:UpdateFirewallPolicyChangeProtection'

-> UpdateFirewallPolicyChangeProtection

Create a value of UpdateFirewallPolicyChangeProtection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateToken:UpdateFirewallPolicyChangeProtection', updateFirewallPolicyChangeProtection_updateToken - An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

$sel:firewallArn:UpdateFirewallPolicyChangeProtection', updateFirewallPolicyChangeProtection_firewallArn - The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

$sel:firewallName:UpdateFirewallPolicyChangeProtection', updateFirewallPolicyChangeProtection_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

$sel:firewallPolicyChangeProtection:UpdateFirewallPolicyChangeProtection', updateFirewallPolicyChangeProtection_firewallPolicyChangeProtection - A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

Request Lenses

updateFirewallPolicyChangeProtection_updateToken :: Lens' UpdateFirewallPolicyChangeProtection (Maybe Text)

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

updateFirewallPolicyChangeProtection_firewallArn :: Lens' UpdateFirewallPolicyChangeProtection (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

updateFirewallPolicyChangeProtection_firewallName :: Lens' UpdateFirewallPolicyChangeProtection (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

updateFirewallPolicyChangeProtection_firewallPolicyChangeProtection :: Lens' UpdateFirewallPolicyChangeProtection Bool

A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

Destructuring the Response

data UpdateFirewallPolicyChangeProtectionResponse

See: newUpdateFirewallPolicyChangeProtectionResponse smart constructor.

Constructors

UpdateFirewallPolicyChangeProtectionResponse'

Fields

updateToken :: Maybe Text

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

firewallPolicyChangeProtection :: Maybe Bool

A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

httpStatus :: Int

The response's http status code.

newUpdateFirewallPolicyChangeProtectionResponse

Arguments

:: Int
$sel:httpStatus:UpdateFirewallPolicyChangeProtectionResponse'

-> UpdateFirewallPolicyChangeProtectionResponse

Create a value of UpdateFirewallPolicyChangeProtectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateToken:UpdateFirewallPolicyChangeProtection', updateFirewallPolicyChangeProtectionResponse_updateToken - An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

$sel:firewallArn:UpdateFirewallPolicyChangeProtection', updateFirewallPolicyChangeProtectionResponse_firewallArn - The Amazon Resource Name (ARN) of the firewall.

$sel:firewallPolicyChangeProtection:UpdateFirewallPolicyChangeProtection', updateFirewallPolicyChangeProtectionResponse_firewallPolicyChangeProtection - A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

$sel:firewallName:UpdateFirewallPolicyChangeProtection', updateFirewallPolicyChangeProtectionResponse_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

$sel:httpStatus:UpdateFirewallPolicyChangeProtectionResponse', updateFirewallPolicyChangeProtectionResponse_httpStatus - The response's http status code.

Response Lenses

updateFirewallPolicyChangeProtectionResponse_updateToken :: Lens' UpdateFirewallPolicyChangeProtectionResponse (Maybe Text)

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

updateFirewallPolicyChangeProtectionResponse_firewallArn :: Lens' UpdateFirewallPolicyChangeProtectionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

updateFirewallPolicyChangeProtectionResponse_firewallPolicyChangeProtection :: Lens' UpdateFirewallPolicyChangeProtectionResponse (Maybe Bool)

A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

updateFirewallPolicyChangeProtectionResponse_firewallName :: Lens' UpdateFirewallPolicyChangeProtectionResponse (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

updateFirewallPolicyChangeProtectionResponse_httpStatus :: Lens' UpdateFirewallPolicyChangeProtectionResponse Int

The response's http status code.