Amazonka.NetworkManager.AssociateTransitGatewayConnectPeer

Description

Associates a transit gateway Connect peer with a device, and optionally, with a link. If you specify a link, it must be associated with the specified device.

You can only associate transit gateway Connect peers that have been created on a transit gateway that's registered in your global network.

You cannot associate a transit gateway Connect peer with more than one device and link.

Creating a Request

data AssociateTransitGatewayConnectPeer

See: newAssociateTransitGatewayConnectPeer smart constructor.

Constructors

AssociateTransitGatewayConnectPeer'

Fields

linkId :: Maybe Text

The ID of the link.

globalNetworkId :: Text

The ID of the global network.

transitGatewayConnectPeerArn :: Text

The Amazon Resource Name (ARN) of the Connect peer.

deviceId :: Text

The ID of the device.

newAssociateTransitGatewayConnectPeer

Arguments

:: Text
$sel:globalNetworkId:AssociateTransitGatewayConnectPeer'

-> Text
$sel:transitGatewayConnectPeerArn:AssociateTransitGatewayConnectPeer'

-> Text
$sel:deviceId:AssociateTransitGatewayConnectPeer'

-> AssociateTransitGatewayConnectPeer

Create a value of AssociateTransitGatewayConnectPeer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:linkId:AssociateTransitGatewayConnectPeer', associateTransitGatewayConnectPeer_linkId - The ID of the link.

$sel:globalNetworkId:AssociateTransitGatewayConnectPeer', associateTransitGatewayConnectPeer_globalNetworkId - The ID of the global network.

$sel:transitGatewayConnectPeerArn:AssociateTransitGatewayConnectPeer', associateTransitGatewayConnectPeer_transitGatewayConnectPeerArn - The Amazon Resource Name (ARN) of the Connect peer.

$sel:deviceId:AssociateTransitGatewayConnectPeer', associateTransitGatewayConnectPeer_deviceId - The ID of the device.

Request Lenses

associateTransitGatewayConnectPeer_linkId :: Lens' AssociateTransitGatewayConnectPeer (Maybe Text)

The ID of the link.

associateTransitGatewayConnectPeer_globalNetworkId :: Lens' AssociateTransitGatewayConnectPeer Text

The ID of the global network.

associateTransitGatewayConnectPeer_transitGatewayConnectPeerArn :: Lens' AssociateTransitGatewayConnectPeer Text

The Amazon Resource Name (ARN) of the Connect peer.

associateTransitGatewayConnectPeer_deviceId :: Lens' AssociateTransitGatewayConnectPeer Text

The ID of the device.

Destructuring the Response

data AssociateTransitGatewayConnectPeerResponse

See: newAssociateTransitGatewayConnectPeerResponse smart constructor.

Constructors

AssociateTransitGatewayConnectPeerResponse'

Fields

transitGatewayConnectPeerAssociation :: Maybe TransitGatewayConnectPeerAssociation

The transit gateway Connect peer association.

httpStatus :: Int

The response's http status code.

newAssociateTransitGatewayConnectPeerResponse

Arguments

:: Int
$sel:httpStatus:AssociateTransitGatewayConnectPeerResponse'

-> AssociateTransitGatewayConnectPeerResponse

Create a value of AssociateTransitGatewayConnectPeerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transitGatewayConnectPeerAssociation:AssociateTransitGatewayConnectPeerResponse', associateTransitGatewayConnectPeerResponse_transitGatewayConnectPeerAssociation - The transit gateway Connect peer association.

$sel:httpStatus:AssociateTransitGatewayConnectPeerResponse', associateTransitGatewayConnectPeerResponse_httpStatus - The response's http status code.

Response Lenses

associateTransitGatewayConnectPeerResponse_transitGatewayConnectPeerAssociation :: Lens' AssociateTransitGatewayConnectPeerResponse (Maybe TransitGatewayConnectPeerAssociation)

The transit gateway Connect peer association.

associateTransitGatewayConnectPeerResponse_httpStatus :: Lens' AssociateTransitGatewayConnectPeerResponse Int

The response's http status code.