Amazonka.NetworkManager.CreateLink

Description

Creates a new link for a specified site.

Creating a Request

data CreateLink

See: newCreateLink smart constructor.

Constructors

CreateLink'

Fields

type' :: Maybe Text

The type of the link.

Constraints: Cannot include the following characters: | \ ^

Length Constraints: Maximum length of 128 characters.

description :: Maybe Text

A description of the link.

Length Constraints: Maximum length of 256 characters.

provider :: Maybe Text

The provider of the link.

Constraints: Cannot include the following characters: | \ ^

Length Constraints: Maximum length of 128 characters.

tags :: Maybe [Tag]

The tags to apply to the resource during creation.

globalNetworkId :: Text

The ID of the global network.

bandwidth :: Bandwidth

The upload speed and download speed in Mbps.

siteId :: Text

The ID of the site.

newCreateLink

Arguments

:: Text
$sel:globalNetworkId:CreateLink'

-> Bandwidth
$sel:bandwidth:CreateLink'

-> Text
$sel:siteId:CreateLink'

-> CreateLink

Create a value of CreateLink with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':CreateLink', createLink_type - The type of the link.

Constraints: Cannot include the following characters: | \ ^

Length Constraints: Maximum length of 128 characters.

$sel:description:CreateLink', createLink_description - A description of the link.

Length Constraints: Maximum length of 256 characters.

$sel:provider:CreateLink', createLink_provider - The provider of the link.

Constraints: Cannot include the following characters: | \ ^

Length Constraints: Maximum length of 128 characters.

$sel:tags:CreateLink', createLink_tags - The tags to apply to the resource during creation.

$sel:globalNetworkId:CreateLink', createLink_globalNetworkId - The ID of the global network.

$sel:bandwidth:CreateLink', createLink_bandwidth - The upload speed and download speed in Mbps.

$sel:siteId:CreateLink', createLink_siteId - The ID of the site.

Request Lenses

createLink_type :: Lens' CreateLink (Maybe Text)

The type of the link.

Constraints: Cannot include the following characters: | \ ^

Length Constraints: Maximum length of 128 characters.

createLink_description :: Lens' CreateLink (Maybe Text)

A description of the link.

Length Constraints: Maximum length of 256 characters.

createLink_provider :: Lens' CreateLink (Maybe Text)

The provider of the link.

Constraints: Cannot include the following characters: | \ ^

Length Constraints: Maximum length of 128 characters.

createLink_tags :: Lens' CreateLink (Maybe [Tag])

The tags to apply to the resource during creation.

createLink_globalNetworkId :: Lens' CreateLink Text

The ID of the global network.

createLink_bandwidth :: Lens' CreateLink Bandwidth

The upload speed and download speed in Mbps.

createLink_siteId :: Lens' CreateLink Text

The ID of the site.

Destructuring the Response

data CreateLinkResponse

See: newCreateLinkResponse smart constructor.

Constructors

CreateLinkResponse'

Fields

link :: Maybe Link

Information about the link.

httpStatus :: Int

The response's http status code.

newCreateLinkResponse

Arguments

:: Int
$sel:httpStatus:CreateLinkResponse'

-> CreateLinkResponse

Create a value of CreateLinkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:link:CreateLinkResponse', createLinkResponse_link - Information about the link.

$sel:httpStatus:CreateLinkResponse', createLinkResponse_httpStatus - The response's http status code.

Response Lenses

createLinkResponse_link :: Lens' CreateLinkResponse (Maybe Link)

Information about the link.

createLinkResponse_httpStatus :: Lens' CreateLinkResponse Int

The response's http status code.