Amazonka.NetworkManager.GetConnections

Description

Gets information about one or more of your connections in a global network.

This operation returns paginated results.

Creating a Request

data GetConnections

See: newGetConnections smart constructor.

Constructors

GetConnections'

Fields

connectionIds :: Maybe [Text]

One or more connection IDs.

nextToken :: Maybe Text

The token for the next page of results.

deviceId :: Maybe Text

The ID of the device.

maxResults :: Maybe Natural

The maximum number of results to return.

globalNetworkId :: Text

The ID of the global network.

newGetConnections

Arguments

:: Text
$sel:globalNetworkId:GetConnections'

-> GetConnections

Create a value of GetConnections with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionIds:GetConnections', getConnections_connectionIds - One or more connection IDs.

$sel:nextToken:GetConnections', getConnections_nextToken - The token for the next page of results.

$sel:deviceId:GetConnections', getConnections_deviceId - The ID of the device.

$sel:maxResults:GetConnections', getConnections_maxResults - The maximum number of results to return.

$sel:globalNetworkId:GetConnections', getConnections_globalNetworkId - The ID of the global network.

Request Lenses

getConnections_connectionIds :: Lens' GetConnections (Maybe [Text])

One or more connection IDs.

getConnections_nextToken :: Lens' GetConnections (Maybe Text)

The token for the next page of results.

getConnections_deviceId :: Lens' GetConnections (Maybe Text)

The ID of the device.

getConnections_maxResults :: Lens' GetConnections (Maybe Natural)

The maximum number of results to return.

getConnections_globalNetworkId :: Lens' GetConnections Text

The ID of the global network.

Destructuring the Response

data GetConnectionsResponse

See: newGetConnectionsResponse smart constructor.

Constructors

GetConnectionsResponse'

Fields

connections :: Maybe [Connection]

Information about the connections.

nextToken :: Maybe Text

The token to use for the next page of results.

httpStatus :: Int

The response's http status code.

newGetConnectionsResponse

Arguments

:: Int
$sel:httpStatus:GetConnectionsResponse'

-> GetConnectionsResponse

Create a value of GetConnectionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connections:GetConnectionsResponse', getConnectionsResponse_connections - Information about the connections.

$sel:nextToken:GetConnections', getConnectionsResponse_nextToken - The token to use for the next page of results.

$sel:httpStatus:GetConnectionsResponse', getConnectionsResponse_httpStatus - The response's http status code.

Response Lenses

getConnectionsResponse_connections :: Lens' GetConnectionsResponse (Maybe [Connection])

Information about the connections.

getConnectionsResponse_nextToken :: Lens' GetConnectionsResponse (Maybe Text)

The token to use for the next page of results.

getConnectionsResponse_httpStatus :: Lens' GetConnectionsResponse Int

The response's http status code.