Amazonka.NetworkManager.Types

Description

Service Configuration

defaultService :: Service

API version 2019-07-05 of the Amazon Network Manager SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

There was a conflict processing the request. Updating or deleting the resource can cause an inconsistent state.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A service limit was exceeded.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The request has failed due to an internal error.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource could not be found.

ConnectionState

newtype ConnectionState

Constructors

ConnectionState'

Fields

fromConnectionState :: Text

Bundled Patterns

pattern ConnectionState_AVAILABLE :: ConnectionState

pattern ConnectionState_DELETING :: ConnectionState

pattern ConnectionState_PENDING :: ConnectionState

pattern ConnectionState_UPDATING :: ConnectionState

CustomerGatewayAssociationState

newtype CustomerGatewayAssociationState

Constructors

CustomerGatewayAssociationState'

Fields

fromCustomerGatewayAssociationState :: Text

Bundled Patterns

pattern CustomerGatewayAssociationState_AVAILABLE :: CustomerGatewayAssociationState

pattern CustomerGatewayAssociationState_DELETED :: CustomerGatewayAssociationState

pattern CustomerGatewayAssociationState_DELETING :: CustomerGatewayAssociationState

pattern CustomerGatewayAssociationState_PENDING :: CustomerGatewayAssociationState

DeviceState

newtype DeviceState

Constructors

DeviceState'

Fields

fromDeviceState :: Text

Bundled Patterns

pattern DeviceState_AVAILABLE :: DeviceState

pattern DeviceState_DELETING :: DeviceState

pattern DeviceState_PENDING :: DeviceState

pattern DeviceState_UPDATING :: DeviceState

GlobalNetworkState

newtype GlobalNetworkState

Constructors

GlobalNetworkState'

Fields

fromGlobalNetworkState :: Text

Bundled Patterns

pattern GlobalNetworkState_AVAILABLE :: GlobalNetworkState

pattern GlobalNetworkState_DELETING :: GlobalNetworkState

pattern GlobalNetworkState_PENDING :: GlobalNetworkState

pattern GlobalNetworkState_UPDATING :: GlobalNetworkState

LinkAssociationState

newtype LinkAssociationState

Constructors

LinkAssociationState'

Fields

fromLinkAssociationState :: Text

Bundled Patterns

pattern LinkAssociationState_AVAILABLE :: LinkAssociationState

pattern LinkAssociationState_DELETED :: LinkAssociationState

pattern LinkAssociationState_DELETING :: LinkAssociationState

pattern LinkAssociationState_PENDING :: LinkAssociationState

LinkState

newtype LinkState

Constructors

LinkState'

Fields

fromLinkState :: Text

Bundled Patterns

pattern LinkState_AVAILABLE :: LinkState

pattern LinkState_DELETING :: LinkState

pattern LinkState_PENDING :: LinkState

pattern LinkState_UPDATING :: LinkState

SiteState

newtype SiteState

Constructors

SiteState'

Fields

fromSiteState :: Text

Bundled Patterns

pattern SiteState_AVAILABLE :: SiteState

pattern SiteState_DELETING :: SiteState

pattern SiteState_PENDING :: SiteState

pattern SiteState_UPDATING :: SiteState

TransitGatewayConnectPeerAssociationState

newtype TransitGatewayConnectPeerAssociationState

Constructors

TransitGatewayConnectPeerAssociationState'

Fields

fromTransitGatewayConnectPeerAssociationState :: Text

Bundled Patterns

pattern TransitGatewayConnectPeerAssociationState_AVAILABLE :: TransitGatewayConnectPeerAssociationState

pattern TransitGatewayConnectPeerAssociationState_DELETED :: TransitGatewayConnectPeerAssociationState

pattern TransitGatewayConnectPeerAssociationState_DELETING :: TransitGatewayConnectPeerAssociationState

pattern TransitGatewayConnectPeerAssociationState_PENDING :: TransitGatewayConnectPeerAssociationState

TransitGatewayRegistrationState

newtype TransitGatewayRegistrationState

Constructors

TransitGatewayRegistrationState'

Fields

fromTransitGatewayRegistrationState :: Text

Bundled Patterns

pattern TransitGatewayRegistrationState_AVAILABLE :: TransitGatewayRegistrationState

pattern TransitGatewayRegistrationState_DELETED :: TransitGatewayRegistrationState

pattern TransitGatewayRegistrationState_DELETING :: TransitGatewayRegistrationState

pattern TransitGatewayRegistrationState_FAILED :: TransitGatewayRegistrationState

pattern TransitGatewayRegistrationState_PENDING :: TransitGatewayRegistrationState

AWSLocation

data AWSLocation

Specifies a location in AWS.

See: newAWSLocation smart constructor.

Constructors

AWSLocation'

Fields

zone :: Maybe Text

The Zone the device is located in. This can be the ID of an Availability Zone, Local Zone, Wavelength Zone, or an Outpost.

subnetArn :: Maybe Text

The Amazon Resource Name (ARN) of the subnet the device is located in.

newAWSLocation :: AWSLocation

Create a value of AWSLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:zone:AWSLocation', aWSLocation_zone - The Zone the device is located in. This can be the ID of an Availability Zone, Local Zone, Wavelength Zone, or an Outpost.

$sel:subnetArn:AWSLocation', aWSLocation_subnetArn - The Amazon Resource Name (ARN) of the subnet the device is located in.

aWSLocation_zone :: Lens' AWSLocation (Maybe Text)

The Zone the device is located in. This can be the ID of an Availability Zone, Local Zone, Wavelength Zone, or an Outpost.

aWSLocation_subnetArn :: Lens' AWSLocation (Maybe Text)

The Amazon Resource Name (ARN) of the subnet the device is located in.

Bandwidth

data Bandwidth

Describes bandwidth information.

See: newBandwidth smart constructor.

Constructors

Bandwidth'

Fields

downloadSpeed :: Maybe Int

Download speed in Mbps.

uploadSpeed :: Maybe Int

Upload speed in Mbps.

newBandwidth :: Bandwidth

Create a value of Bandwidth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:downloadSpeed:Bandwidth', bandwidth_downloadSpeed - Download speed in Mbps.

$sel:uploadSpeed:Bandwidth', bandwidth_uploadSpeed - Upload speed in Mbps.

bandwidth_downloadSpeed :: Lens' Bandwidth (Maybe Int)

Download speed in Mbps.

bandwidth_uploadSpeed :: Lens' Bandwidth (Maybe Int)

Upload speed in Mbps.

Connection

data Connection

Describes a connection.

See: newConnection smart constructor.

Constructors

Connection'

Fields

state :: Maybe ConnectionState

The state of the connection.

createdAt :: Maybe POSIX

The date and time that the connection was created.

globalNetworkId :: Maybe Text

The ID of the global network.

connectionId :: Maybe Text

The ID of the connection.

connectedDeviceId :: Maybe Text

The ID of the second device in the connection.

connectedLinkId :: Maybe Text

The ID of the link for the second device in the connection.

linkId :: Maybe Text

The ID of the link for the first device in the connection.

deviceId :: Maybe Text

The ID of the first device in the connection.

connectionArn :: Maybe Text

The Amazon Resource Name (ARN) of the connection.

description :: Maybe Text

The description of the connection.

tags :: Maybe [Tag]

The tags for the connection.

newConnection :: Connection

Create a value of Connection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Connection', connection_state - The state of the connection.

$sel:createdAt:Connection', connection_createdAt - The date and time that the connection was created.

$sel:globalNetworkId:Connection', connection_globalNetworkId - The ID of the global network.

$sel:connectionId:Connection', connection_connectionId - The ID of the connection.

$sel:connectedDeviceId:Connection', connection_connectedDeviceId - The ID of the second device in the connection.

$sel:connectedLinkId:Connection', connection_connectedLinkId - The ID of the link for the second device in the connection.

$sel:linkId:Connection', connection_linkId - The ID of the link for the first device in the connection.

$sel:deviceId:Connection', connection_deviceId - The ID of the first device in the connection.

$sel:connectionArn:Connection', connection_connectionArn - The Amazon Resource Name (ARN) of the connection.

$sel:description:Connection', connection_description - The description of the connection.

$sel:tags:Connection', connection_tags - The tags for the connection.

connection_state :: Lens' Connection (Maybe ConnectionState)

The state of the connection.

connection_createdAt :: Lens' Connection (Maybe UTCTime)

The date and time that the connection was created.

connection_globalNetworkId :: Lens' Connection (Maybe Text)

The ID of the global network.

connection_connectionId :: Lens' Connection (Maybe Text)

The ID of the connection.

connection_connectedDeviceId :: Lens' Connection (Maybe Text)

The ID of the second device in the connection.

connection_connectedLinkId :: Lens' Connection (Maybe Text)

The ID of the link for the second device in the connection.

connection_linkId :: Lens' Connection (Maybe Text)

The ID of the link for the first device in the connection.

connection_deviceId :: Lens' Connection (Maybe Text)

The ID of the first device in the connection.

connection_connectionArn :: Lens' Connection (Maybe Text)

The Amazon Resource Name (ARN) of the connection.

connection_description :: Lens' Connection (Maybe Text)

The description of the connection.

connection_tags :: Lens' Connection (Maybe [Tag])

The tags for the connection.

CustomerGatewayAssociation

data CustomerGatewayAssociation

Describes the association between a customer gateway, a device, and a link.

See: newCustomerGatewayAssociation smart constructor.

Constructors

CustomerGatewayAssociation'

Fields

state :: Maybe CustomerGatewayAssociationState

The association state.

globalNetworkId :: Maybe Text

The ID of the global network.

linkId :: Maybe Text

The ID of the link.

deviceId :: Maybe Text

The ID of the device.

customerGatewayArn :: Maybe Text

The Amazon Resource Name (ARN) of the customer gateway.

newCustomerGatewayAssociation :: CustomerGatewayAssociation

Create a value of CustomerGatewayAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:CustomerGatewayAssociation', customerGatewayAssociation_state - The association state.

$sel:globalNetworkId:CustomerGatewayAssociation', customerGatewayAssociation_globalNetworkId - The ID of the global network.

$sel:linkId:CustomerGatewayAssociation', customerGatewayAssociation_linkId - The ID of the link.

$sel:deviceId:CustomerGatewayAssociation', customerGatewayAssociation_deviceId - The ID of the device.

$sel:customerGatewayArn:CustomerGatewayAssociation', customerGatewayAssociation_customerGatewayArn - The Amazon Resource Name (ARN) of the customer gateway.

customerGatewayAssociation_state :: Lens' CustomerGatewayAssociation (Maybe CustomerGatewayAssociationState)

The association state.

customerGatewayAssociation_globalNetworkId :: Lens' CustomerGatewayAssociation (Maybe Text)

The ID of the global network.

customerGatewayAssociation_linkId :: Lens' CustomerGatewayAssociation (Maybe Text)

The ID of the link.

customerGatewayAssociation_deviceId :: Lens' CustomerGatewayAssociation (Maybe Text)

The ID of the device.

customerGatewayAssociation_customerGatewayArn :: Lens' CustomerGatewayAssociation (Maybe Text)

The Amazon Resource Name (ARN) of the customer gateway.

Device

data Device

Describes a device.

See: newDevice smart constructor.

Constructors

Device'

Fields

vendor :: Maybe Text

The device vendor.

state :: Maybe DeviceState

The device state.

location :: Maybe (Sensitive Location)

The site location.

createdAt :: Maybe POSIX

The date and time that the site was created.

deviceArn :: Maybe Text

The Amazon Resource Name (ARN) of the device.

aWSLocation :: Maybe AWSLocation

The AWS location of the device.

globalNetworkId :: Maybe Text

The ID of the global network.

model :: Maybe Text

The device model.

deviceId :: Maybe Text

The ID of the device.

type' :: Maybe Text

The device type.

serialNumber :: Maybe Text

The device serial number.

siteId :: Maybe Text

The site ID.

description :: Maybe Text

The description of the device.

tags :: Maybe [Tag]

The tags for the device.

newDevice :: Device

Create a value of Device with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vendor:Device', device_vendor - The device vendor.

$sel:state:Device', device_state - The device state.

$sel:location:Device', device_location - The site location.

$sel:createdAt:Device', device_createdAt - The date and time that the site was created.

$sel:deviceArn:Device', device_deviceArn - The Amazon Resource Name (ARN) of the device.

$sel:aWSLocation:Device', device_aWSLocation - The AWS location of the device.

$sel:globalNetworkId:Device', device_globalNetworkId - The ID of the global network.

$sel:model:Device', device_model - The device model.

$sel:deviceId:Device', device_deviceId - The ID of the device.

$sel:type':Device', device_type - The device type.

$sel:serialNumber:Device', device_serialNumber - The device serial number.

$sel:siteId:Device', device_siteId - The site ID.

$sel:description:Device', device_description - The description of the device.

$sel:tags:Device', device_tags - The tags for the device.

device_vendor :: Lens' Device (Maybe Text)

The device vendor.

device_state :: Lens' Device (Maybe DeviceState)

The device state.

device_location :: Lens' Device (Maybe Location)

The site location.

device_createdAt :: Lens' Device (Maybe UTCTime)

The date and time that the site was created.

device_deviceArn :: Lens' Device (Maybe Text)

The Amazon Resource Name (ARN) of the device.

device_aWSLocation :: Lens' Device (Maybe AWSLocation)

The AWS location of the device.

device_globalNetworkId :: Lens' Device (Maybe Text)

The ID of the global network.

device_model :: Lens' Device (Maybe Text)

The device model.

device_deviceId :: Lens' Device (Maybe Text)

The ID of the device.

device_type :: Lens' Device (Maybe Text)

The device type.

device_serialNumber :: Lens' Device (Maybe Text)

The device serial number.

device_siteId :: Lens' Device (Maybe Text)

The site ID.

device_description :: Lens' Device (Maybe Text)

The description of the device.

device_tags :: Lens' Device (Maybe [Tag])

The tags for the device.

GlobalNetwork

data GlobalNetwork

Describes a global network.

See: newGlobalNetwork smart constructor.

Constructors

GlobalNetwork'

Fields

state :: Maybe GlobalNetworkState

The state of the global network.

createdAt :: Maybe POSIX

The date and time that the global network was created.

globalNetworkArn :: Maybe Text

The Amazon Resource Name (ARN) of the global network.

globalNetworkId :: Maybe Text

The ID of the global network.

description :: Maybe Text

The description of the global network.

tags :: Maybe [Tag]

The tags for the global network.

newGlobalNetwork :: GlobalNetwork

Create a value of GlobalNetwork with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:GlobalNetwork', globalNetwork_state - The state of the global network.

$sel:createdAt:GlobalNetwork', globalNetwork_createdAt - The date and time that the global network was created.

$sel:globalNetworkArn:GlobalNetwork', globalNetwork_globalNetworkArn - The Amazon Resource Name (ARN) of the global network.

$sel:globalNetworkId:GlobalNetwork', globalNetwork_globalNetworkId - The ID of the global network.

$sel:description:GlobalNetwork', globalNetwork_description - The description of the global network.

$sel:tags:GlobalNetwork', globalNetwork_tags - The tags for the global network.

globalNetwork_state :: Lens' GlobalNetwork (Maybe GlobalNetworkState)

The state of the global network.

globalNetwork_createdAt :: Lens' GlobalNetwork (Maybe UTCTime)

The date and time that the global network was created.

globalNetwork_globalNetworkArn :: Lens' GlobalNetwork (Maybe Text)

The Amazon Resource Name (ARN) of the global network.

globalNetwork_globalNetworkId :: Lens' GlobalNetwork (Maybe Text)

The ID of the global network.

globalNetwork_description :: Lens' GlobalNetwork (Maybe Text)

The description of the global network.

globalNetwork_tags :: Lens' GlobalNetwork (Maybe [Tag])

The tags for the global network.

Link

data Link

Describes a link.

See: newLink smart constructor.

Constructors

Link'

Fields

state :: Maybe LinkState

The state of the link.

linkArn :: Maybe Text

The Amazon Resource Name (ARN) of the link.

createdAt :: Maybe POSIX

The date and time that the link was created.

globalNetworkId :: Maybe Text

The ID of the global network.

bandwidth :: Maybe Bandwidth

The bandwidth for the link.

linkId :: Maybe Text

The ID of the link.

type' :: Maybe Text

The type of the link.

siteId :: Maybe Text

The ID of the site.

description :: Maybe Text

The description of the link.

provider :: Maybe Text

The provider of the link.

tags :: Maybe [Tag]

The tags for the link.

newLink :: Link

Create a value of Link with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Link', link_state - The state of the link.

$sel:linkArn:Link', link_linkArn - The Amazon Resource Name (ARN) of the link.

$sel:createdAt:Link', link_createdAt - The date and time that the link was created.

$sel:globalNetworkId:Link', link_globalNetworkId - The ID of the global network.

$sel:bandwidth:Link', link_bandwidth - The bandwidth for the link.

$sel:linkId:Link', link_linkId - The ID of the link.

$sel:type':Link', link_type - The type of the link.

$sel:siteId:Link', link_siteId - The ID of the site.

$sel:description:Link', link_description - The description of the link.

$sel:provider:Link', link_provider - The provider of the link.

$sel:tags:Link', link_tags - The tags for the link.

link_state :: Lens' Link (Maybe LinkState)

The state of the link.

link_linkArn :: Lens' Link (Maybe Text)

The Amazon Resource Name (ARN) of the link.

link_createdAt :: Lens' Link (Maybe UTCTime)

The date and time that the link was created.

link_globalNetworkId :: Lens' Link (Maybe Text)

The ID of the global network.

link_bandwidth :: Lens' Link (Maybe Bandwidth)

The bandwidth for the link.

link_linkId :: Lens' Link (Maybe Text)

The ID of the link.

link_type :: Lens' Link (Maybe Text)

The type of the link.

link_siteId :: Lens' Link (Maybe Text)

The ID of the site.

link_description :: Lens' Link (Maybe Text)

The description of the link.

link_provider :: Lens' Link (Maybe Text)

The provider of the link.

link_tags :: Lens' Link (Maybe [Tag])

The tags for the link.

LinkAssociation

data LinkAssociation

Describes the association between a device and a link.

See: newLinkAssociation smart constructor.

Constructors

LinkAssociation'

Fields

globalNetworkId :: Maybe Text

The ID of the global network.

linkId :: Maybe Text

The ID of the link.

deviceId :: Maybe Text

The device ID for the link association.

linkAssociationState :: Maybe LinkAssociationState

The state of the association.

newLinkAssociation :: LinkAssociation

Create a value of LinkAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalNetworkId:LinkAssociation', linkAssociation_globalNetworkId - The ID of the global network.

$sel:linkId:LinkAssociation', linkAssociation_linkId - The ID of the link.

$sel:deviceId:LinkAssociation', linkAssociation_deviceId - The device ID for the link association.

$sel:linkAssociationState:LinkAssociation', linkAssociation_linkAssociationState - The state of the association.

linkAssociation_globalNetworkId :: Lens' LinkAssociation (Maybe Text)

The ID of the global network.

linkAssociation_linkId :: Lens' LinkAssociation (Maybe Text)

The ID of the link.

linkAssociation_deviceId :: Lens' LinkAssociation (Maybe Text)

The device ID for the link association.

linkAssociation_linkAssociationState :: Lens' LinkAssociation (Maybe LinkAssociationState)

The state of the association.

Location

data Location

Describes a location.

See: newLocation smart constructor.

Constructors

Location'

Fields

latitude :: Maybe Text

The latitude.

address :: Maybe Text

The physical address.

longitude :: Maybe Text

The longitude.

newLocation :: Location

Create a value of Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latitude:Location', location_latitude - The latitude.

$sel:address:Location', location_address - The physical address.

$sel:longitude:Location', location_longitude - The longitude.

location_latitude :: Lens' Location (Maybe Text)

The latitude.

location_address :: Lens' Location (Maybe Text)

The physical address.

location_longitude :: Lens' Location (Maybe Text)

The longitude.

Site

data Site

Describes a site.

See: newSite smart constructor.

Constructors

Site'

Fields

state :: Maybe SiteState

The state of the site.

location :: Maybe (Sensitive Location)

The location of the site.

createdAt :: Maybe POSIX

The date and time that the site was created.

globalNetworkId :: Maybe Text

The ID of the global network.

siteId :: Maybe Text

The ID of the site.

siteArn :: Maybe Text

The Amazon Resource Name (ARN) of the site.

description :: Maybe Text

The description of the site.

tags :: Maybe [Tag]

The tags for the site.

newSite :: Site

Create a value of Site with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Site', site_state - The state of the site.

$sel:location:Site', site_location - The location of the site.

$sel:createdAt:Site', site_createdAt - The date and time that the site was created.

$sel:globalNetworkId:Site', site_globalNetworkId - The ID of the global network.

$sel:siteId:Site', site_siteId - The ID of the site.

$sel:siteArn:Site', site_siteArn - The Amazon Resource Name (ARN) of the site.

$sel:description:Site', site_description - The description of the site.

$sel:tags:Site', site_tags - The tags for the site.

site_state :: Lens' Site (Maybe SiteState)

The state of the site.

site_location :: Lens' Site (Maybe Location)

The location of the site.

site_createdAt :: Lens' Site (Maybe UTCTime)

The date and time that the site was created.

site_globalNetworkId :: Lens' Site (Maybe Text)

The ID of the global network.

site_siteId :: Lens' Site (Maybe Text)

The ID of the site.

site_siteArn :: Lens' Site (Maybe Text)

The Amazon Resource Name (ARN) of the site.

site_description :: Lens' Site (Maybe Text)

The description of the site.

site_tags :: Lens' Site (Maybe [Tag])

The tags for the site.

Tag

data Tag

Describes a tag.

See: newTag smart constructor.

Constructors

Tag'

Fields

value :: Maybe Text

The tag value.

Length Constraints: Maximum length of 256 characters.

key :: Maybe Text

The tag key.

Length Constraints: Maximum length of 128 characters.

newTag :: Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - The tag value.

Length Constraints: Maximum length of 256 characters.

$sel:key:Tag', tag_key - The tag key.

Length Constraints: Maximum length of 128 characters.

tag_value :: Lens' Tag (Maybe Text)

The tag value.

Length Constraints: Maximum length of 256 characters.

tag_key :: Lens' Tag (Maybe Text)

The tag key.

Length Constraints: Maximum length of 128 characters.

TransitGatewayConnectPeerAssociation

data TransitGatewayConnectPeerAssociation

Describes a transit gateway Connect peer association.

See: newTransitGatewayConnectPeerAssociation smart constructor.

Constructors

TransitGatewayConnectPeerAssociation'

Fields

state :: Maybe TransitGatewayConnectPeerAssociationState

The state of the association.

globalNetworkId :: Maybe Text

The ID of the global network.

transitGatewayConnectPeerArn :: Maybe Text

The Amazon Resource Name (ARN) of the transit gateway Connect peer.

linkId :: Maybe Text

The ID of the link.

deviceId :: Maybe Text

The ID of the device.

newTransitGatewayConnectPeerAssociation :: TransitGatewayConnectPeerAssociation

Create a value of TransitGatewayConnectPeerAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:TransitGatewayConnectPeerAssociation', transitGatewayConnectPeerAssociation_state - The state of the association.

$sel:globalNetworkId:TransitGatewayConnectPeerAssociation', transitGatewayConnectPeerAssociation_globalNetworkId - The ID of the global network.

$sel:transitGatewayConnectPeerArn:TransitGatewayConnectPeerAssociation', transitGatewayConnectPeerAssociation_transitGatewayConnectPeerArn - The Amazon Resource Name (ARN) of the transit gateway Connect peer.

$sel:linkId:TransitGatewayConnectPeerAssociation', transitGatewayConnectPeerAssociation_linkId - The ID of the link.

$sel:deviceId:TransitGatewayConnectPeerAssociation', transitGatewayConnectPeerAssociation_deviceId - The ID of the device.

transitGatewayConnectPeerAssociation_state :: Lens' TransitGatewayConnectPeerAssociation (Maybe TransitGatewayConnectPeerAssociationState)

The state of the association.

transitGatewayConnectPeerAssociation_globalNetworkId :: Lens' TransitGatewayConnectPeerAssociation (Maybe Text)

The ID of the global network.

transitGatewayConnectPeerAssociation_transitGatewayConnectPeerArn :: Lens' TransitGatewayConnectPeerAssociation (Maybe Text)

The Amazon Resource Name (ARN) of the transit gateway Connect peer.

transitGatewayConnectPeerAssociation_linkId :: Lens' TransitGatewayConnectPeerAssociation (Maybe Text)

The ID of the link.

transitGatewayConnectPeerAssociation_deviceId :: Lens' TransitGatewayConnectPeerAssociation (Maybe Text)

The ID of the device.

TransitGatewayRegistration

data TransitGatewayRegistration

Describes the registration of a transit gateway to a global network.

See: newTransitGatewayRegistration smart constructor.

Constructors

TransitGatewayRegistration'

Fields

state :: Maybe TransitGatewayRegistrationStateReason

The state of the transit gateway registration.

globalNetworkId :: Maybe Text

The ID of the global network.

transitGatewayArn :: Maybe Text

The Amazon Resource Name (ARN) of the transit gateway.

newTransitGatewayRegistration :: TransitGatewayRegistration

Create a value of TransitGatewayRegistration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:TransitGatewayRegistration', transitGatewayRegistration_state - The state of the transit gateway registration.

$sel:globalNetworkId:TransitGatewayRegistration', transitGatewayRegistration_globalNetworkId - The ID of the global network.

$sel:transitGatewayArn:TransitGatewayRegistration', transitGatewayRegistration_transitGatewayArn - The Amazon Resource Name (ARN) of the transit gateway.

transitGatewayRegistration_state :: Lens' TransitGatewayRegistration (Maybe TransitGatewayRegistrationStateReason)

The state of the transit gateway registration.

transitGatewayRegistration_globalNetworkId :: Lens' TransitGatewayRegistration (Maybe Text)

The ID of the global network.

transitGatewayRegistration_transitGatewayArn :: Lens' TransitGatewayRegistration (Maybe Text)

The Amazon Resource Name (ARN) of the transit gateway.

TransitGatewayRegistrationStateReason

data TransitGatewayRegistrationStateReason

Describes the status of a transit gateway registration.

See: newTransitGatewayRegistrationStateReason smart constructor.

Constructors

TransitGatewayRegistrationStateReason'

Fields

code :: Maybe TransitGatewayRegistrationState

The code for the state reason.

message :: Maybe Text

The message for the state reason.

newTransitGatewayRegistrationStateReason :: TransitGatewayRegistrationStateReason

Create a value of TransitGatewayRegistrationStateReason with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:code:TransitGatewayRegistrationStateReason', transitGatewayRegistrationStateReason_code - The code for the state reason.

$sel:message:TransitGatewayRegistrationStateReason', transitGatewayRegistrationStateReason_message - The message for the state reason.

transitGatewayRegistrationStateReason_code :: Lens' TransitGatewayRegistrationStateReason (Maybe TransitGatewayRegistrationState)

The code for the state reason.

transitGatewayRegistrationStateReason_message :: Lens' TransitGatewayRegistrationStateReason (Maybe Text)

The message for the state reason.