Amazonka.NetworkManager.Types.Site

Description

Documentation

data Site

Describes a site.

See: newSite smart constructor.

Constructors

Site'

Fields

state :: Maybe SiteState

The state of the site.

location :: Maybe (Sensitive Location)

The location of the site.

createdAt :: Maybe POSIX

The date and time that the site was created.

globalNetworkId :: Maybe Text

The ID of the global network.

siteId :: Maybe Text

The ID of the site.

siteArn :: Maybe Text

The Amazon Resource Name (ARN) of the site.

description :: Maybe Text

The description of the site.

tags :: Maybe [Tag]

The tags for the site.

newSite :: Site

Create a value of Site with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Site', site_state - The state of the site.

$sel:location:Site', site_location - The location of the site.

$sel:createdAt:Site', site_createdAt - The date and time that the site was created.

$sel:globalNetworkId:Site', site_globalNetworkId - The ID of the global network.

$sel:siteId:Site', site_siteId - The ID of the site.

$sel:siteArn:Site', site_siteArn - The Amazon Resource Name (ARN) of the site.

$sel:description:Site', site_description - The description of the site.

$sel:tags:Site', site_tags - The tags for the site.

site_state :: Lens' Site (Maybe SiteState)

The state of the site.

site_location :: Lens' Site (Maybe Location)

The location of the site.

site_createdAt :: Lens' Site (Maybe UTCTime)

The date and time that the site was created.

site_globalNetworkId :: Lens' Site (Maybe Text)

The ID of the global network.

site_siteId :: Lens' Site (Maybe Text)

The ID of the site.

site_siteArn :: Lens' Site (Maybe Text)

The Amazon Resource Name (ARN) of the site.

site_description :: Lens' Site (Maybe Text)

The description of the site.

site_tags :: Lens' Site (Maybe [Tag])

The tags for the site.