Amazonka.Nimble.CreateLaunchProfile

Description

Create a launch profile.

Creating a Request

data CreateLaunchProfile

A collection of launch profiles.

See: newCreateLaunchProfile smart constructor.

Constructors

CreateLaunchProfile'

Fields

clientToken :: Maybe Text

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

description :: Maybe Text

The description.

tags :: Maybe (HashMap Text Text)

A collection of labels, in the form of key:value pairs, that apply to this resource.

ec2SubnetIds :: [Text]

studioComponentIds :: NonEmpty Text

Unique identifiers for a collection of studio components that can be used with this launch profile.

studioId :: Text

The studio ID.

launchProfileProtocolVersions :: [Text]

The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

name :: Text

The name for the launch profile.

streamConfiguration :: StreamConfigurationCreate

A configuration for a streaming session.

newCreateLaunchProfile

Arguments

:: NonEmpty Text
$sel:studioComponentIds:CreateLaunchProfile'

-> Text
$sel:studioId:CreateLaunchProfile'

-> Text
$sel:name:CreateLaunchProfile'

-> StreamConfigurationCreate
$sel:streamConfiguration:CreateLaunchProfile'

-> CreateLaunchProfile

Create a value of CreateLaunchProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateLaunchProfile', createLaunchProfile_clientToken - To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

$sel:description:CreateLaunchProfile', createLaunchProfile_description - The description.

$sel:tags:CreateLaunchProfile', createLaunchProfile_tags - A collection of labels, in the form of key:value pairs, that apply to this resource.

$sel:ec2SubnetIds:CreateLaunchProfile', createLaunchProfile_ec2SubnetIds -

$sel:studioComponentIds:CreateLaunchProfile', createLaunchProfile_studioComponentIds - Unique identifiers for a collection of studio components that can be used with this launch profile.

$sel:studioId:CreateLaunchProfile', createLaunchProfile_studioId - The studio ID.

$sel:launchProfileProtocolVersions:CreateLaunchProfile', createLaunchProfile_launchProfileProtocolVersions - The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

$sel:name:CreateLaunchProfile', createLaunchProfile_name - The name for the launch profile.

$sel:streamConfiguration:CreateLaunchProfile', createLaunchProfile_streamConfiguration - A configuration for a streaming session.

Request Lenses

createLaunchProfile_clientToken :: Lens' CreateLaunchProfile (Maybe Text)

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

createLaunchProfile_description :: Lens' CreateLaunchProfile (Maybe Text)

The description.

createLaunchProfile_tags :: Lens' CreateLaunchProfile (Maybe (HashMap Text Text))

A collection of labels, in the form of key:value pairs, that apply to this resource.

createLaunchProfile_ec2SubnetIds :: Lens' CreateLaunchProfile [Text]

createLaunchProfile_studioComponentIds :: Lens' CreateLaunchProfile (NonEmpty Text)

Unique identifiers for a collection of studio components that can be used with this launch profile.

createLaunchProfile_studioId :: Lens' CreateLaunchProfile Text

The studio ID.

createLaunchProfile_launchProfileProtocolVersions :: Lens' CreateLaunchProfile [Text]

The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

createLaunchProfile_name :: Lens' CreateLaunchProfile Text

The name for the launch profile.

createLaunchProfile_streamConfiguration :: Lens' CreateLaunchProfile StreamConfigurationCreate

A configuration for a streaming session.

Destructuring the Response

data CreateLaunchProfileResponse

See: newCreateLaunchProfileResponse smart constructor.

Constructors

CreateLaunchProfileResponse'

Fields

launchProfile :: Maybe LaunchProfile

The launch profile.

httpStatus :: Int

The response's http status code.

newCreateLaunchProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateLaunchProfileResponse'

-> CreateLaunchProfileResponse

Create a value of CreateLaunchProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchProfile:CreateLaunchProfileResponse', createLaunchProfileResponse_launchProfile - The launch profile.

$sel:httpStatus:CreateLaunchProfileResponse', createLaunchProfileResponse_httpStatus - The response's http status code.

Response Lenses

createLaunchProfileResponse_launchProfile :: Lens' CreateLaunchProfileResponse (Maybe LaunchProfile)

The launch profile.

createLaunchProfileResponse_httpStatus :: Lens' CreateLaunchProfileResponse Int

The response's http status code.