Amazonka.Nimble.CreateStreamingImage

Description

Creates a streaming image resource in a studio.

Creating a Request

data CreateStreamingImage

A collection of streaming images.

See: newCreateStreamingImage smart constructor.

Constructors

CreateStreamingImage'

Fields

clientToken :: Maybe Text

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

description :: Maybe Text

A human-readable description of the streaming image.

tags :: Maybe (HashMap Text Text)

A collection of labels, in the form of key:value pairs, that apply to this resource.

studioId :: Text

The studio ID.

name :: Text

A friendly name for a streaming image resource.

ec2ImageId :: Text

The ID of an EC2 machine image with which to create this streaming image.

newCreateStreamingImage

Arguments

:: Text
$sel:studioId:CreateStreamingImage'

-> Text
$sel:name:CreateStreamingImage'

-> Text
$sel:ec2ImageId:CreateStreamingImage'

-> CreateStreamingImage

Create a value of CreateStreamingImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateStreamingImage', createStreamingImage_clientToken - To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

$sel:description:CreateStreamingImage', createStreamingImage_description - A human-readable description of the streaming image.

$sel:tags:CreateStreamingImage', createStreamingImage_tags - A collection of labels, in the form of key:value pairs, that apply to this resource.

$sel:studioId:CreateStreamingImage', createStreamingImage_studioId - The studio ID.

$sel:name:CreateStreamingImage', createStreamingImage_name - A friendly name for a streaming image resource.

$sel:ec2ImageId:CreateStreamingImage', createStreamingImage_ec2ImageId - The ID of an EC2 machine image with which to create this streaming image.

Request Lenses

createStreamingImage_clientToken :: Lens' CreateStreamingImage (Maybe Text)

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

createStreamingImage_description :: Lens' CreateStreamingImage (Maybe Text)

A human-readable description of the streaming image.

createStreamingImage_tags :: Lens' CreateStreamingImage (Maybe (HashMap Text Text))

A collection of labels, in the form of key:value pairs, that apply to this resource.

createStreamingImage_studioId :: Lens' CreateStreamingImage Text

The studio ID.

createStreamingImage_name :: Lens' CreateStreamingImage Text

A friendly name for a streaming image resource.

createStreamingImage_ec2ImageId :: Lens' CreateStreamingImage Text

The ID of an EC2 machine image with which to create this streaming image.

Destructuring the Response

data CreateStreamingImageResponse

See: newCreateStreamingImageResponse smart constructor.

Constructors

CreateStreamingImageResponse'

Fields

streamingImage :: Maybe StreamingImage

httpStatus :: Int

The response's http status code.

newCreateStreamingImageResponse

Arguments

:: Int
$sel:httpStatus:CreateStreamingImageResponse'

-> CreateStreamingImageResponse

Create a value of CreateStreamingImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamingImage:CreateStreamingImageResponse', createStreamingImageResponse_streamingImage -

$sel:httpStatus:CreateStreamingImageResponse', createStreamingImageResponse_httpStatus - The response's http status code.

Response Lenses

createStreamingImageResponse_streamingImage :: Lens' CreateStreamingImageResponse (Maybe StreamingImage)

createStreamingImageResponse_httpStatus :: Lens' CreateStreamingImageResponse Int

The response's http status code.