Amazonka.Nimble.CreateStudioComponent

Description

Creates a studio component resource.

Creating a Request

data CreateStudioComponent

The studio components.

See: newCreateStudioComponent smart constructor.

Constructors

CreateStudioComponent'

Fields

initializationScripts :: Maybe [StudioComponentInitializationScript]

Initialization scripts for studio components.

clientToken :: Maybe Text

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

ec2SecurityGroupIds :: Maybe (NonEmpty Text)

The EC2 security groups that control access to the studio component.

subtype :: Maybe StudioComponentSubtype

The specific subtype of a studio component.

scriptParameters :: Maybe [ScriptParameterKeyValue]

Parameters for the studio component scripts.

configuration :: Maybe StudioComponentConfiguration

The configuration of the studio component, based on component type.

description :: Maybe Text

The description.

tags :: Maybe (HashMap Text Text)

A collection of labels, in the form of key:value pairs, that apply to this resource.

studioId :: Text

The studio ID.

name :: Text

The name for the studio component.

type' :: StudioComponentType

The type of the studio component.

newCreateStudioComponent

Arguments

:: Text
$sel:studioId:CreateStudioComponent'

-> Text
$sel:name:CreateStudioComponent'

-> StudioComponentType
$sel:type':CreateStudioComponent'

-> CreateStudioComponent

Create a value of CreateStudioComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:initializationScripts:CreateStudioComponent', createStudioComponent_initializationScripts - Initialization scripts for studio components.

$sel:clientToken:CreateStudioComponent', createStudioComponent_clientToken - To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

$sel:ec2SecurityGroupIds:CreateStudioComponent', createStudioComponent_ec2SecurityGroupIds - The EC2 security groups that control access to the studio component.

$sel:subtype:CreateStudioComponent', createStudioComponent_subtype - The specific subtype of a studio component.

$sel:scriptParameters:CreateStudioComponent', createStudioComponent_scriptParameters - Parameters for the studio component scripts.

$sel:configuration:CreateStudioComponent', createStudioComponent_configuration - The configuration of the studio component, based on component type.

$sel:description:CreateStudioComponent', createStudioComponent_description - The description.

$sel:tags:CreateStudioComponent', createStudioComponent_tags - A collection of labels, in the form of key:value pairs, that apply to this resource.

$sel:studioId:CreateStudioComponent', createStudioComponent_studioId - The studio ID.

$sel:name:CreateStudioComponent', createStudioComponent_name - The name for the studio component.

$sel:type':CreateStudioComponent', createStudioComponent_type - The type of the studio component.

Request Lenses

createStudioComponent_initializationScripts :: Lens' CreateStudioComponent (Maybe [StudioComponentInitializationScript])

Initialization scripts for studio components.

createStudioComponent_clientToken :: Lens' CreateStudioComponent (Maybe Text)

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

createStudioComponent_ec2SecurityGroupIds :: Lens' CreateStudioComponent (Maybe (NonEmpty Text))

The EC2 security groups that control access to the studio component.

createStudioComponent_subtype :: Lens' CreateStudioComponent (Maybe StudioComponentSubtype)

The specific subtype of a studio component.

createStudioComponent_scriptParameters :: Lens' CreateStudioComponent (Maybe [ScriptParameterKeyValue])

Parameters for the studio component scripts.

createStudioComponent_configuration :: Lens' CreateStudioComponent (Maybe StudioComponentConfiguration)

The configuration of the studio component, based on component type.

createStudioComponent_description :: Lens' CreateStudioComponent (Maybe Text)

The description.

createStudioComponent_tags :: Lens' CreateStudioComponent (Maybe (HashMap Text Text))

A collection of labels, in the form of key:value pairs, that apply to this resource.

createStudioComponent_studioId :: Lens' CreateStudioComponent Text

The studio ID.

createStudioComponent_name :: Lens' CreateStudioComponent Text

The name for the studio component.

createStudioComponent_type :: Lens' CreateStudioComponent StudioComponentType

The type of the studio component.

Destructuring the Response

data CreateStudioComponentResponse

See: newCreateStudioComponentResponse smart constructor.

Constructors

CreateStudioComponentResponse'

Fields

studioComponent :: Maybe StudioComponent

Information about the studio component.

httpStatus :: Int

The response's http status code.

newCreateStudioComponentResponse

Arguments

:: Int
$sel:httpStatus:CreateStudioComponentResponse'

-> CreateStudioComponentResponse

Create a value of CreateStudioComponentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioComponent:CreateStudioComponentResponse', createStudioComponentResponse_studioComponent - Information about the studio component.

$sel:httpStatus:CreateStudioComponentResponse', createStudioComponentResponse_httpStatus - The response's http status code.

Response Lenses

createStudioComponentResponse_studioComponent :: Lens' CreateStudioComponentResponse (Maybe StudioComponent)

Information about the studio component.

createStudioComponentResponse_httpStatus :: Lens' CreateStudioComponentResponse Int

The response's http status code.