Amazonka.Nimble.GetLaunchProfile

Description

Get a launch profile.

Creating a Request

data GetLaunchProfile

See: newGetLaunchProfile smart constructor.

Constructors

GetLaunchProfile'

Fields

studioId :: Text

The studio ID.

launchProfileId :: Text

The launch profile ID.

newGetLaunchProfile

Arguments

:: Text
$sel:studioId:GetLaunchProfile'

-> Text
$sel:launchProfileId:GetLaunchProfile'

-> GetLaunchProfile

Create a value of GetLaunchProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioId:GetLaunchProfile', getLaunchProfile_studioId - The studio ID.

$sel:launchProfileId:GetLaunchProfile', getLaunchProfile_launchProfileId - The launch profile ID.

Request Lenses

getLaunchProfile_studioId :: Lens' GetLaunchProfile Text

The studio ID.

getLaunchProfile_launchProfileId :: Lens' GetLaunchProfile Text

The launch profile ID.

Destructuring the Response

data GetLaunchProfileResponse

See: newGetLaunchProfileResponse smart constructor.

Constructors

GetLaunchProfileResponse'

Fields

launchProfile :: Maybe LaunchProfile

The launch profile.

httpStatus :: Int

The response's http status code.

newGetLaunchProfileResponse

Arguments

:: Int
$sel:httpStatus:GetLaunchProfileResponse'

-> GetLaunchProfileResponse

Create a value of GetLaunchProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchProfile:GetLaunchProfileResponse', getLaunchProfileResponse_launchProfile - The launch profile.

$sel:httpStatus:GetLaunchProfileResponse', getLaunchProfileResponse_httpStatus - The response's http status code.

Response Lenses

getLaunchProfileResponse_launchProfile :: Lens' GetLaunchProfileResponse (Maybe LaunchProfile)

The launch profile.

getLaunchProfileResponse_httpStatus :: Lens' GetLaunchProfileResponse Int

The response's http status code.