Amazonka.Nimble.GetLaunchProfileDetails

Description

Launch profile details include the launch profile resource and summary information of resources that are used by, or available to, the launch profile. This includes the name and description of all studio components used by the launch profiles, and the name and description of streaming images that can be used with this launch profile.

Creating a Request

data GetLaunchProfileDetails

See: newGetLaunchProfileDetails smart constructor.

Constructors

GetLaunchProfileDetails'

Fields

studioId :: Text

The studio ID.

launchProfileId :: Text

The launch profile ID.

newGetLaunchProfileDetails

Arguments

:: Text
$sel:studioId:GetLaunchProfileDetails'

-> Text
$sel:launchProfileId:GetLaunchProfileDetails'

-> GetLaunchProfileDetails

Create a value of GetLaunchProfileDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioId:GetLaunchProfileDetails', getLaunchProfileDetails_studioId - The studio ID.

$sel:launchProfileId:GetLaunchProfileDetails', getLaunchProfileDetails_launchProfileId - The launch profile ID.

Request Lenses

getLaunchProfileDetails_studioId :: Lens' GetLaunchProfileDetails Text

The studio ID.

getLaunchProfileDetails_launchProfileId :: Lens' GetLaunchProfileDetails Text

The launch profile ID.

Destructuring the Response

data GetLaunchProfileDetailsResponse

See: newGetLaunchProfileDetailsResponse smart constructor.

Constructors

GetLaunchProfileDetailsResponse'

Fields

streamingImages :: Maybe [StreamingImage]

A collection of streaming images.

launchProfile :: Maybe LaunchProfile

The launch profile.

studioComponentSummaries :: Maybe [StudioComponentSummary]

A collection of studio component summaries.

httpStatus :: Int

The response's http status code.

newGetLaunchProfileDetailsResponse

Arguments

:: Int
$sel:httpStatus:GetLaunchProfileDetailsResponse'

-> GetLaunchProfileDetailsResponse

Create a value of GetLaunchProfileDetailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamingImages:GetLaunchProfileDetailsResponse', getLaunchProfileDetailsResponse_streamingImages - A collection of streaming images.

$sel:launchProfile:GetLaunchProfileDetailsResponse', getLaunchProfileDetailsResponse_launchProfile - The launch profile.

$sel:studioComponentSummaries:GetLaunchProfileDetailsResponse', getLaunchProfileDetailsResponse_studioComponentSummaries - A collection of studio component summaries.

$sel:httpStatus:GetLaunchProfileDetailsResponse', getLaunchProfileDetailsResponse_httpStatus - The response's http status code.

Response Lenses

getLaunchProfileDetailsResponse_streamingImages :: Lens' GetLaunchProfileDetailsResponse (Maybe [StreamingImage])

A collection of streaming images.

getLaunchProfileDetailsResponse_launchProfile :: Lens' GetLaunchProfileDetailsResponse (Maybe LaunchProfile)

The launch profile.

getLaunchProfileDetailsResponse_studioComponentSummaries :: Lens' GetLaunchProfileDetailsResponse (Maybe [StudioComponentSummary])

A collection of studio component summaries.

getLaunchProfileDetailsResponse_httpStatus :: Lens' GetLaunchProfileDetailsResponse Int

The response's http status code.