Amazonka.Nimble.GetLaunchProfileInitialization

Description

Get a launch profile initialization.

Creating a Request

data GetLaunchProfileInitialization

See: newGetLaunchProfileInitialization smart constructor.

Constructors

GetLaunchProfileInitialization'

Fields

studioId :: Text

The studio ID.

launchProfileProtocolVersions :: [Text]

A collection of launch profile protocol versions.

launchPurpose :: Text

The launch purpose.

launchProfileId :: Text

The launch profile ID.

platform :: Text

The platform.

newGetLaunchProfileInitialization

Arguments

:: Text
$sel:studioId:GetLaunchProfileInitialization'

-> Text
$sel:launchPurpose:GetLaunchProfileInitialization'

-> Text
$sel:launchProfileId:GetLaunchProfileInitialization'

-> Text
$sel:platform:GetLaunchProfileInitialization'

-> GetLaunchProfileInitialization

Create a value of GetLaunchProfileInitialization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioId:GetLaunchProfileInitialization', getLaunchProfileInitialization_studioId - The studio ID.

$sel:launchProfileProtocolVersions:GetLaunchProfileInitialization', getLaunchProfileInitialization_launchProfileProtocolVersions - A collection of launch profile protocol versions.

$sel:launchPurpose:GetLaunchProfileInitialization', getLaunchProfileInitialization_launchPurpose - The launch purpose.

$sel:launchProfileId:GetLaunchProfileInitialization', getLaunchProfileInitialization_launchProfileId - The launch profile ID.

$sel:platform:GetLaunchProfileInitialization', getLaunchProfileInitialization_platform - The platform.

Request Lenses

getLaunchProfileInitialization_studioId :: Lens' GetLaunchProfileInitialization Text

The studio ID.

getLaunchProfileInitialization_launchProfileProtocolVersions :: Lens' GetLaunchProfileInitialization [Text]

A collection of launch profile protocol versions.

getLaunchProfileInitialization_launchPurpose :: Lens' GetLaunchProfileInitialization Text

The launch purpose.

getLaunchProfileInitialization_launchProfileId :: Lens' GetLaunchProfileInitialization Text

The launch profile ID.

getLaunchProfileInitialization_platform :: Lens' GetLaunchProfileInitialization Text

The platform.

Destructuring the Response

data GetLaunchProfileInitializationResponse

See: newGetLaunchProfileInitializationResponse smart constructor.

Constructors

GetLaunchProfileInitializationResponse'

Fields

launchProfileInitialization :: Maybe LaunchProfileInitialization

The launch profile initialization.

httpStatus :: Int

The response's http status code.

newGetLaunchProfileInitializationResponse

Arguments

:: Int
$sel:httpStatus:GetLaunchProfileInitializationResponse'

-> GetLaunchProfileInitializationResponse

Create a value of GetLaunchProfileInitializationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchProfileInitialization:GetLaunchProfileInitializationResponse', getLaunchProfileInitializationResponse_launchProfileInitialization - The launch profile initialization.

$sel:httpStatus:GetLaunchProfileInitializationResponse', getLaunchProfileInitializationResponse_httpStatus - The response's http status code.

Response Lenses

getLaunchProfileInitializationResponse_launchProfileInitialization :: Lens' GetLaunchProfileInitializationResponse (Maybe LaunchProfileInitialization)

The launch profile initialization.

getLaunchProfileInitializationResponse_httpStatus :: Lens' GetLaunchProfileInitializationResponse Int

The response's http status code.