Amazonka.Nimble.GetStudioComponent

Description

Gets a studio component resource.

Creating a Request

data GetStudioComponent

See: newGetStudioComponent smart constructor.

Constructors

GetStudioComponent'

Fields

studioId :: Text

The studio ID.

studioComponentId :: Text

The studio component ID.

newGetStudioComponent

Arguments

:: Text
$sel:studioId:GetStudioComponent'

-> Text
$sel:studioComponentId:GetStudioComponent'

-> GetStudioComponent

Create a value of GetStudioComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioId:GetStudioComponent', getStudioComponent_studioId - The studio ID.

$sel:studioComponentId:GetStudioComponent', getStudioComponent_studioComponentId - The studio component ID.

Request Lenses

getStudioComponent_studioId :: Lens' GetStudioComponent Text

The studio ID.

getStudioComponent_studioComponentId :: Lens' GetStudioComponent Text

The studio component ID.

Destructuring the Response

data GetStudioComponentResponse

See: newGetStudioComponentResponse smart constructor.

Constructors

GetStudioComponentResponse'

Fields

studioComponent :: Maybe StudioComponent

Information about the studio component.

httpStatus :: Int

The response's http status code.

newGetStudioComponentResponse

Arguments

:: Int
$sel:httpStatus:GetStudioComponentResponse'

-> GetStudioComponentResponse

Create a value of GetStudioComponentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioComponent:GetStudioComponentResponse', getStudioComponentResponse_studioComponent - Information about the studio component.

$sel:httpStatus:GetStudioComponentResponse', getStudioComponentResponse_httpStatus - The response's http status code.

Response Lenses

getStudioComponentResponse_studioComponent :: Lens' GetStudioComponentResponse (Maybe StudioComponent)

Information about the studio component.

getStudioComponentResponse_httpStatus :: Lens' GetStudioComponentResponse Int

The response's http status code.