Amazonka.Nimble.PutStudioMembers

Description

Add/update users with given persona to studio membership.

Creating a Request

data PutStudioMembers

A launch profile membership collection.

See: newPutStudioMembers smart constructor.

Constructors

PutStudioMembers'

Fields

clientToken :: Maybe Text

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

studioId :: Text

The studio ID.

members :: NonEmpty NewStudioMember

A list of members.

identityStoreId :: Text

The ID of the identity store.

newPutStudioMembers

Arguments

:: Text
$sel:studioId:PutStudioMembers'

-> NonEmpty NewStudioMember
$sel:members:PutStudioMembers'

-> Text
$sel:identityStoreId:PutStudioMembers'

-> PutStudioMembers

Create a value of PutStudioMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:PutStudioMembers', putStudioMembers_clientToken - To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

$sel:studioId:PutStudioMembers', putStudioMembers_studioId - The studio ID.

$sel:members:PutStudioMembers', putStudioMembers_members - A list of members.

$sel:identityStoreId:PutStudioMembers', putStudioMembers_identityStoreId - The ID of the identity store.

Request Lenses

putStudioMembers_clientToken :: Lens' PutStudioMembers (Maybe Text)

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

putStudioMembers_studioId :: Lens' PutStudioMembers Text

The studio ID.

putStudioMembers_members :: Lens' PutStudioMembers (NonEmpty NewStudioMember)

A list of members.

putStudioMembers_identityStoreId :: Lens' PutStudioMembers Text

The ID of the identity store.

Destructuring the Response

data PutStudioMembersResponse

See: newPutStudioMembersResponse smart constructor.

Constructors

PutStudioMembersResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutStudioMembersResponse

Arguments

:: Int
$sel:httpStatus:PutStudioMembersResponse'

-> PutStudioMembersResponse

Create a value of PutStudioMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutStudioMembersResponse', putStudioMembersResponse_httpStatus - The response's http status code.

Response Lenses

putStudioMembersResponse_httpStatus :: Lens' PutStudioMembersResponse Int

The response's http status code.