Amazonka.Nimble.Types.EulaAcceptance

Description

Documentation

data EulaAcceptance

See: newEulaAcceptance smart constructor.

Constructors

EulaAcceptance'

Fields

accepteeId :: Maybe Text

The ID of the acceptee.

eulaId :: Maybe Text

The EULA ID.

acceptedAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the EULA was accepted.

acceptedBy :: Maybe Text

The ID of the person who accepted the EULA.

eulaAcceptanceId :: Maybe Text

The EULA acceptance ID.

newEulaAcceptance :: EulaAcceptance

Create a value of EulaAcceptance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accepteeId:EulaAcceptance', eulaAcceptance_accepteeId - The ID of the acceptee.

$sel:eulaId:EulaAcceptance', eulaAcceptance_eulaId - The EULA ID.

$sel:acceptedAt:EulaAcceptance', eulaAcceptance_acceptedAt - The Unix epoch timestamp in seconds for when the EULA was accepted.

$sel:acceptedBy:EulaAcceptance', eulaAcceptance_acceptedBy - The ID of the person who accepted the EULA.

$sel:eulaAcceptanceId:EulaAcceptance', eulaAcceptance_eulaAcceptanceId - The EULA acceptance ID.

eulaAcceptance_accepteeId :: Lens' EulaAcceptance (Maybe Text)

The ID of the acceptee.

eulaAcceptance_eulaId :: Lens' EulaAcceptance (Maybe Text)

The EULA ID.

eulaAcceptance_acceptedAt :: Lens' EulaAcceptance (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the EULA was accepted.

eulaAcceptance_acceptedBy :: Lens' EulaAcceptance (Maybe Text)

The ID of the person who accepted the EULA.

eulaAcceptance_eulaAcceptanceId :: Lens' EulaAcceptance (Maybe Text)

The EULA acceptance ID.