Amazonka.Nimble.Types.LaunchProfile

Description

Documentation

data LaunchProfile

See: newLaunchProfile smart constructor.

Constructors

LaunchProfile'

Fields

state :: Maybe LaunchProfileState

The current state.

arn :: Maybe Text

The ARN of the resource.

createdAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the resource was created.

createdBy :: Maybe Text

The user ID of the user that created the launch profile.

launchProfileId :: Maybe Text

The launch profile ID.

updatedBy :: Maybe Text

The user ID of the user that most recently updated the resource.

launchProfileProtocolVersions :: Maybe [Text]

The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

ec2SubnetIds :: Maybe [Text]

Unique identifiers for a collection of EC2 subnets.

streamConfiguration :: Maybe StreamConfiguration

A configuration for a streaming session.

name :: Maybe Text

A friendly name for the launch profile.

statusMessage :: Maybe Text

The status message for the launch profile.

updatedAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the resource was updated.

description :: Maybe Text

A human-readable description of the launch profile.

tags :: Maybe (HashMap Text Text)

A collection of labels, in the form of key:value pairs, that apply to this resource.

statusCode :: Maybe LaunchProfileStatusCode

The status code.

studioComponentIds :: Maybe (NonEmpty Text)

Unique identifiers for a collection of studio components that can be used with this launch profile.

newLaunchProfile :: LaunchProfile

Create a value of LaunchProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:LaunchProfile', launchProfile_state - The current state.

$sel:arn:LaunchProfile', launchProfile_arn - The ARN of the resource.

$sel:createdAt:LaunchProfile', launchProfile_createdAt - The Unix epoch timestamp in seconds for when the resource was created.

$sel:createdBy:LaunchProfile', launchProfile_createdBy - The user ID of the user that created the launch profile.

$sel:launchProfileId:LaunchProfile', launchProfile_launchProfileId - The launch profile ID.

$sel:updatedBy:LaunchProfile', launchProfile_updatedBy - The user ID of the user that most recently updated the resource.

$sel:launchProfileProtocolVersions:LaunchProfile', launchProfile_launchProfileProtocolVersions - The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

$sel:ec2SubnetIds:LaunchProfile', launchProfile_ec2SubnetIds - Unique identifiers for a collection of EC2 subnets.

$sel:streamConfiguration:LaunchProfile', launchProfile_streamConfiguration - A configuration for a streaming session.

$sel:name:LaunchProfile', launchProfile_name - A friendly name for the launch profile.

$sel:statusMessage:LaunchProfile', launchProfile_statusMessage - The status message for the launch profile.

$sel:updatedAt:LaunchProfile', launchProfile_updatedAt - The Unix epoch timestamp in seconds for when the resource was updated.

$sel:description:LaunchProfile', launchProfile_description - A human-readable description of the launch profile.

$sel:tags:LaunchProfile', launchProfile_tags - A collection of labels, in the form of key:value pairs, that apply to this resource.

$sel:statusCode:LaunchProfile', launchProfile_statusCode - The status code.

$sel:studioComponentIds:LaunchProfile', launchProfile_studioComponentIds - Unique identifiers for a collection of studio components that can be used with this launch profile.

launchProfile_state :: Lens' LaunchProfile (Maybe LaunchProfileState)

The current state.

launchProfile_arn :: Lens' LaunchProfile (Maybe Text)

The ARN of the resource.

launchProfile_createdAt :: Lens' LaunchProfile (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the resource was created.

launchProfile_createdBy :: Lens' LaunchProfile (Maybe Text)

The user ID of the user that created the launch profile.

launchProfile_launchProfileId :: Lens' LaunchProfile (Maybe Text)

The launch profile ID.

launchProfile_updatedBy :: Lens' LaunchProfile (Maybe Text)

The user ID of the user that most recently updated the resource.

launchProfile_launchProfileProtocolVersions :: Lens' LaunchProfile (Maybe [Text])

The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

launchProfile_ec2SubnetIds :: Lens' LaunchProfile (Maybe [Text])

Unique identifiers for a collection of EC2 subnets.

launchProfile_streamConfiguration :: Lens' LaunchProfile (Maybe StreamConfiguration)

A configuration for a streaming session.

launchProfile_name :: Lens' LaunchProfile (Maybe Text)

A friendly name for the launch profile.

launchProfile_statusMessage :: Lens' LaunchProfile (Maybe Text)

The status message for the launch profile.

launchProfile_updatedAt :: Lens' LaunchProfile (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the resource was updated.

launchProfile_description :: Lens' LaunchProfile (Maybe Text)

A human-readable description of the launch profile.

launchProfile_tags :: Lens' LaunchProfile (Maybe (HashMap Text Text))

A collection of labels, in the form of key:value pairs, that apply to this resource.

launchProfile_statusCode :: Lens' LaunchProfile (Maybe LaunchProfileStatusCode)

The status code.

launchProfile_studioComponentIds :: Lens' LaunchProfile (Maybe (NonEmpty Text))

Unique identifiers for a collection of studio components that can be used with this launch profile.