Amazonka.Nimble.Types.StreamConfiguration

Description

Documentation

data StreamConfiguration

A configuration for a streaming session.

See: newStreamConfiguration smart constructor.

Constructors

StreamConfiguration'

Fields

streamingImageIds :: Maybe [Text]

The streaming images that users can select from when launching a streaming session with this launch profile.

maxSessionLengthInMinutes :: Maybe Natural

The length of time, in minutes, that a streaming session can run. After this point, Nimble Studio automatically terminates the session.

clipboardMode :: Maybe StreamingClipboardMode

Enable or disable the use of the system clipboard to copy and paste between the streaming session and streaming client.

ec2InstanceTypes :: Maybe (NonEmpty StreamingInstanceType)

The EC2 instance types that users can select from when launching a streaming session with this launch profile.

newStreamConfiguration :: StreamConfiguration

Create a value of StreamConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamingImageIds:StreamConfiguration', streamConfiguration_streamingImageIds - The streaming images that users can select from when launching a streaming session with this launch profile.

$sel:maxSessionLengthInMinutes:StreamConfiguration', streamConfiguration_maxSessionLengthInMinutes - The length of time, in minutes, that a streaming session can run. After this point, Nimble Studio automatically terminates the session.

$sel:clipboardMode:StreamConfiguration', streamConfiguration_clipboardMode - Enable or disable the use of the system clipboard to copy and paste between the streaming session and streaming client.

$sel:ec2InstanceTypes:StreamConfiguration', streamConfiguration_ec2InstanceTypes - The EC2 instance types that users can select from when launching a streaming session with this launch profile.

streamConfiguration_streamingImageIds :: Lens' StreamConfiguration (Maybe [Text])

The streaming images that users can select from when launching a streaming session with this launch profile.

streamConfiguration_maxSessionLengthInMinutes :: Lens' StreamConfiguration (Maybe Natural)

The length of time, in minutes, that a streaming session can run. After this point, Nimble Studio automatically terminates the session.

streamConfiguration_clipboardMode :: Lens' StreamConfiguration (Maybe StreamingClipboardMode)

Enable or disable the use of the system clipboard to copy and paste between the streaming session and streaming client.

streamConfiguration_ec2InstanceTypes :: Lens' StreamConfiguration (Maybe (NonEmpty StreamingInstanceType))

The EC2 instance types that users can select from when launching a streaming session with this launch profile.