Amazonka.Nimble.Types.StreamingImage

Description

Documentation

data StreamingImage

See: newStreamingImage smart constructor.

Constructors

StreamingImage'

Fields

state :: Maybe StreamingImageState

The current state.

platform :: Maybe Text

The platform of the streaming image, either WINDOWS or LINUX.

arn :: Maybe Text

The ARN of the resource.

streamingImageId :: Maybe Text

The ID of the streaming image.

ec2ImageId :: Maybe Text

The ID of an EC2 machine image with which to create the streaming image.

owner :: Maybe Text

The owner of the streaming image, either the studioId that contains the streaming image, or 'amazon' for images that are provided by Amazon Nimble Studio.

name :: Maybe Text

A friendly name for a streaming image resource.

encryptionConfiguration :: Maybe StreamingImageEncryptionConfiguration

The encryption configuration.

statusMessage :: Maybe Text

The status message for the streaming image.

eulaIds :: Maybe [Text]

The list of EULAs that must be accepted before a Streaming Session can be started using this streaming image.

description :: Maybe Text

A human-readable description of the streaming image.

tags :: Maybe (HashMap Text Text)

A collection of labels, in the form of key:value pairs, that apply to this resource.

statusCode :: Maybe StreamingImageStatusCode

The status code.

newStreamingImage :: StreamingImage

Create a value of StreamingImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:StreamingImage', streamingImage_state - The current state.

$sel:platform:StreamingImage', streamingImage_platform - The platform of the streaming image, either WINDOWS or LINUX.

$sel:arn:StreamingImage', streamingImage_arn - The ARN of the resource.

$sel:streamingImageId:StreamingImage', streamingImage_streamingImageId - The ID of the streaming image.

$sel:ec2ImageId:StreamingImage', streamingImage_ec2ImageId - The ID of an EC2 machine image with which to create the streaming image.

$sel:owner:StreamingImage', streamingImage_owner - The owner of the streaming image, either the studioId that contains the streaming image, or 'amazon' for images that are provided by Amazon Nimble Studio.

$sel:name:StreamingImage', streamingImage_name - A friendly name for a streaming image resource.

$sel:encryptionConfiguration:StreamingImage', streamingImage_encryptionConfiguration - The encryption configuration.

$sel:statusMessage:StreamingImage', streamingImage_statusMessage - The status message for the streaming image.

$sel:eulaIds:StreamingImage', streamingImage_eulaIds - The list of EULAs that must be accepted before a Streaming Session can be started using this streaming image.

$sel:description:StreamingImage', streamingImage_description - A human-readable description of the streaming image.

$sel:tags:StreamingImage', streamingImage_tags - A collection of labels, in the form of key:value pairs, that apply to this resource.

$sel:statusCode:StreamingImage', streamingImage_statusCode - The status code.

streamingImage_state :: Lens' StreamingImage (Maybe StreamingImageState)

The current state.

streamingImage_platform :: Lens' StreamingImage (Maybe Text)

The platform of the streaming image, either WINDOWS or LINUX.

streamingImage_arn :: Lens' StreamingImage (Maybe Text)

The ARN of the resource.

streamingImage_streamingImageId :: Lens' StreamingImage (Maybe Text)

The ID of the streaming image.

streamingImage_ec2ImageId :: Lens' StreamingImage (Maybe Text)

The ID of an EC2 machine image with which to create the streaming image.

streamingImage_owner :: Lens' StreamingImage (Maybe Text)

The owner of the streaming image, either the studioId that contains the streaming image, or 'amazon' for images that are provided by Amazon Nimble Studio.

streamingImage_name :: Lens' StreamingImage (Maybe Text)

A friendly name for a streaming image resource.

streamingImage_encryptionConfiguration :: Lens' StreamingImage (Maybe StreamingImageEncryptionConfiguration)

The encryption configuration.

streamingImage_statusMessage :: Lens' StreamingImage (Maybe Text)

The status message for the streaming image.

streamingImage_eulaIds :: Lens' StreamingImage (Maybe [Text])

The list of EULAs that must be accepted before a Streaming Session can be started using this streaming image.

streamingImage_description :: Lens' StreamingImage (Maybe Text)

A human-readable description of the streaming image.

streamingImage_tags :: Lens' StreamingImage (Maybe (HashMap Text Text))

A collection of labels, in the form of key:value pairs, that apply to this resource.

streamingImage_statusCode :: Lens' StreamingImage (Maybe StreamingImageStatusCode)

The status code.