Amazonka.Nimble.UpdateLaunchProfile

Description

Update a launch profile.

Creating a Request

data UpdateLaunchProfile

The launch profile ID.

See: newUpdateLaunchProfile smart constructor.

Constructors

UpdateLaunchProfile'

Fields

clientToken :: Maybe Text

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

launchProfileProtocolVersions :: Maybe [Text]

The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

streamConfiguration :: Maybe StreamConfigurationCreate

A configuration for a streaming session.

name :: Maybe Text

The name for the launch profile.

description :: Maybe Text

The description.

studioComponentIds :: Maybe (NonEmpty Text)

Unique identifiers for a collection of studio components that can be used with this launch profile.

studioId :: Text

The studio ID.

launchProfileId :: Text

The launch profile ID.

newUpdateLaunchProfile

Arguments

:: Text
$sel:studioId:UpdateLaunchProfile'

-> Text
$sel:launchProfileId:UpdateLaunchProfile'

-> UpdateLaunchProfile

Create a value of UpdateLaunchProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateLaunchProfile', updateLaunchProfile_clientToken - To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

$sel:launchProfileProtocolVersions:UpdateLaunchProfile', updateLaunchProfile_launchProfileProtocolVersions - The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

$sel:streamConfiguration:UpdateLaunchProfile', updateLaunchProfile_streamConfiguration - A configuration for a streaming session.

$sel:name:UpdateLaunchProfile', updateLaunchProfile_name - The name for the launch profile.

$sel:description:UpdateLaunchProfile', updateLaunchProfile_description - The description.

$sel:studioComponentIds:UpdateLaunchProfile', updateLaunchProfile_studioComponentIds - Unique identifiers for a collection of studio components that can be used with this launch profile.

$sel:studioId:UpdateLaunchProfile', updateLaunchProfile_studioId - The studio ID.

$sel:launchProfileId:UpdateLaunchProfile', updateLaunchProfile_launchProfileId - The launch profile ID.

Request Lenses

updateLaunchProfile_clientToken :: Lens' UpdateLaunchProfile (Maybe Text)

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

updateLaunchProfile_launchProfileProtocolVersions :: Lens' UpdateLaunchProfile (Maybe [Text])

The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

updateLaunchProfile_streamConfiguration :: Lens' UpdateLaunchProfile (Maybe StreamConfigurationCreate)

A configuration for a streaming session.

updateLaunchProfile_name :: Lens' UpdateLaunchProfile (Maybe Text)

The name for the launch profile.

updateLaunchProfile_description :: Lens' UpdateLaunchProfile (Maybe Text)

The description.

updateLaunchProfile_studioComponentIds :: Lens' UpdateLaunchProfile (Maybe (NonEmpty Text))

Unique identifiers for a collection of studio components that can be used with this launch profile.

updateLaunchProfile_studioId :: Lens' UpdateLaunchProfile Text

The studio ID.

updateLaunchProfile_launchProfileId :: Lens' UpdateLaunchProfile Text

The launch profile ID.

Destructuring the Response

data UpdateLaunchProfileResponse

See: newUpdateLaunchProfileResponse smart constructor.

Constructors

UpdateLaunchProfileResponse'

Fields

launchProfile :: Maybe LaunchProfile

The launch profile.

httpStatus :: Int

The response's http status code.

newUpdateLaunchProfileResponse

Arguments

:: Int
$sel:httpStatus:UpdateLaunchProfileResponse'

-> UpdateLaunchProfileResponse

Create a value of UpdateLaunchProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchProfile:UpdateLaunchProfileResponse', updateLaunchProfileResponse_launchProfile - The launch profile.

$sel:httpStatus:UpdateLaunchProfileResponse', updateLaunchProfileResponse_httpStatus - The response's http status code.

Response Lenses

updateLaunchProfileResponse_launchProfile :: Lens' UpdateLaunchProfileResponse (Maybe LaunchProfile)

The launch profile.

updateLaunchProfileResponse_httpStatus :: Lens' UpdateLaunchProfileResponse Int

The response's http status code.