Amazonka.OpenSearch.CreateDomain

Description

Creates a new Amazon OpenSearch Service domain. For more information, see Creating and managing Amazon OpenSearch Service domains in the Amazon OpenSearch Service Developer Guide.

Creating a Request

data CreateDomain

See: newCreateDomain smart constructor.

Constructors

CreateDomain'

Fields

eBSOptions :: Maybe EBSOptions

Options to enable, disable, and specify the type and size of EBS storage volumes.

engineVersion :: Maybe Text

String of format Elasticsearch_X.Y or OpenSearch_X.Y to specify the engine version for the Amazon OpenSearch Service domain. For example, "OpenSearch_1.0" or "Elasticsearch_7.9". For more information, see Creating and managing Amazon OpenSearch Service domains .

nodeToNodeEncryptionOptions :: Maybe NodeToNodeEncryptionOptions

Node-to-node encryption options.

accessPolicies :: Maybe Text

IAM access policy as a JSON-formatted string.

autoTuneOptions :: Maybe AutoTuneOptionsInput

Specifies Auto-Tune options.

logPublishingOptions :: Maybe (HashMap LogType LogPublishingOption)

Map of LogType and LogPublishingOption, each containing options to publish a given type of OpenSearch log.

clusterConfig :: Maybe ClusterConfig

Configuration options for a domain. Specifies the instance type and number of instances in the domain.

advancedSecurityOptions :: Maybe AdvancedSecurityOptionsInput

Specifies advanced security options.

tagList :: Maybe [Tag]

A list of Tag added during domain creation.

snapshotOptions :: Maybe SnapshotOptions

Option to set time, in UTC format, of the daily automated snapshot. Default value is 0 hours.

cognitoOptions :: Maybe CognitoOptions

Options to specify the Cognito user and identity pools for OpenSearch Dashboards authentication. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

encryptionAtRestOptions :: Maybe EncryptionAtRestOptions

Options for encryption of data at rest.

vPCOptions :: Maybe VPCOptions

Options to specify the subnets and security groups for a VPC endpoint. For more information, see Launching your Amazon OpenSearch Service domains using a VPC .

domainEndpointOptions :: Maybe DomainEndpointOptions

Options to specify configurations that will be applied to the domain endpoint.

advancedOptions :: Maybe (HashMap Text Text)

Option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Advanced cluster parameters for more information.

domainName :: Text

The name of the Amazon OpenSearch Service domain you're creating. Domain names are unique across the domains owned by an account within an AWS region. Domain names must start with a lowercase letter and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

newCreateDomain

Arguments

:: Text
$sel:domainName:CreateDomain'

-> CreateDomain

Create a value of CreateDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eBSOptions:CreateDomain', createDomain_eBSOptions - Options to enable, disable, and specify the type and size of EBS storage volumes.

$sel:engineVersion:CreateDomain', createDomain_engineVersion - String of format Elasticsearch_X.Y or OpenSearch_X.Y to specify the engine version for the Amazon OpenSearch Service domain. For example, "OpenSearch_1.0" or "Elasticsearch_7.9". For more information, see Creating and managing Amazon OpenSearch Service domains .

$sel:nodeToNodeEncryptionOptions:CreateDomain', createDomain_nodeToNodeEncryptionOptions - Node-to-node encryption options.

$sel:accessPolicies:CreateDomain', createDomain_accessPolicies - IAM access policy as a JSON-formatted string.

$sel:autoTuneOptions:CreateDomain', createDomain_autoTuneOptions - Specifies Auto-Tune options.

$sel:logPublishingOptions:CreateDomain', createDomain_logPublishingOptions - Map of LogType and LogPublishingOption, each containing options to publish a given type of OpenSearch log.

$sel:clusterConfig:CreateDomain', createDomain_clusterConfig - Configuration options for a domain. Specifies the instance type and number of instances in the domain.

$sel:advancedSecurityOptions:CreateDomain', createDomain_advancedSecurityOptions - Specifies advanced security options.

$sel:tagList:CreateDomain', createDomain_tagList - A list of Tag added during domain creation.

$sel:snapshotOptions:CreateDomain', createDomain_snapshotOptions - Option to set time, in UTC format, of the daily automated snapshot. Default value is 0 hours.

$sel:cognitoOptions:CreateDomain', createDomain_cognitoOptions - Options to specify the Cognito user and identity pools for OpenSearch Dashboards authentication. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

$sel:encryptionAtRestOptions:CreateDomain', createDomain_encryptionAtRestOptions - Options for encryption of data at rest.

$sel:vPCOptions:CreateDomain', createDomain_vPCOptions - Options to specify the subnets and security groups for a VPC endpoint. For more information, see Launching your Amazon OpenSearch Service domains using a VPC .

$sel:domainEndpointOptions:CreateDomain', createDomain_domainEndpointOptions - Options to specify configurations that will be applied to the domain endpoint.

$sel:advancedOptions:CreateDomain', createDomain_advancedOptions - Option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Advanced cluster parameters for more information.

$sel:domainName:CreateDomain', createDomain_domainName - The name of the Amazon OpenSearch Service domain you're creating. Domain names are unique across the domains owned by an account within an AWS region. Domain names must start with a lowercase letter and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

Request Lenses

createDomain_eBSOptions :: Lens' CreateDomain (Maybe EBSOptions)

Options to enable, disable, and specify the type and size of EBS storage volumes.

createDomain_engineVersion :: Lens' CreateDomain (Maybe Text)

String of format Elasticsearch_X.Y or OpenSearch_X.Y to specify the engine version for the Amazon OpenSearch Service domain. For example, "OpenSearch_1.0" or "Elasticsearch_7.9". For more information, see Creating and managing Amazon OpenSearch Service domains .

createDomain_nodeToNodeEncryptionOptions :: Lens' CreateDomain (Maybe NodeToNodeEncryptionOptions)

Node-to-node encryption options.

createDomain_accessPolicies :: Lens' CreateDomain (Maybe Text)

IAM access policy as a JSON-formatted string.

createDomain_autoTuneOptions :: Lens' CreateDomain (Maybe AutoTuneOptionsInput)

Specifies Auto-Tune options.

createDomain_logPublishingOptions :: Lens' CreateDomain (Maybe (HashMap LogType LogPublishingOption))

Map of LogType and LogPublishingOption, each containing options to publish a given type of OpenSearch log.

createDomain_clusterConfig :: Lens' CreateDomain (Maybe ClusterConfig)

Configuration options for a domain. Specifies the instance type and number of instances in the domain.

createDomain_advancedSecurityOptions :: Lens' CreateDomain (Maybe AdvancedSecurityOptionsInput)

Specifies advanced security options.

createDomain_tagList :: Lens' CreateDomain (Maybe [Tag])

A list of Tag added during domain creation.

createDomain_snapshotOptions :: Lens' CreateDomain (Maybe SnapshotOptions)

Option to set time, in UTC format, of the daily automated snapshot. Default value is 0 hours.

createDomain_cognitoOptions :: Lens' CreateDomain (Maybe CognitoOptions)

Options to specify the Cognito user and identity pools for OpenSearch Dashboards authentication. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

createDomain_encryptionAtRestOptions :: Lens' CreateDomain (Maybe EncryptionAtRestOptions)

Options for encryption of data at rest.

createDomain_vPCOptions :: Lens' CreateDomain (Maybe VPCOptions)

Options to specify the subnets and security groups for a VPC endpoint. For more information, see Launching your Amazon OpenSearch Service domains using a VPC .

createDomain_domainEndpointOptions :: Lens' CreateDomain (Maybe DomainEndpointOptions)

Options to specify configurations that will be applied to the domain endpoint.

createDomain_advancedOptions :: Lens' CreateDomain (Maybe (HashMap Text Text))

Option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Advanced cluster parameters for more information.

createDomain_domainName :: Lens' CreateDomain Text

The name of the Amazon OpenSearch Service domain you're creating. Domain names are unique across the domains owned by an account within an AWS region. Domain names must start with a lowercase letter and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

Destructuring the Response

data CreateDomainResponse

The result of a CreateDomain operation. Contains the status of the newly created Amazon OpenSearch Service domain.

See: newCreateDomainResponse smart constructor.

Constructors

CreateDomainResponse'

Fields

domainStatus :: Maybe DomainStatus

The status of the newly created domain.

httpStatus :: Int

The response's http status code.

newCreateDomainResponse

Arguments

:: Int
$sel:httpStatus:CreateDomainResponse'

-> CreateDomainResponse

Create a value of CreateDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainStatus:CreateDomainResponse', createDomainResponse_domainStatus - The status of the newly created domain.

$sel:httpStatus:CreateDomainResponse', createDomainResponse_httpStatus - The response's http status code.

Response Lenses

createDomainResponse_domainStatus :: Lens' CreateDomainResponse (Maybe DomainStatus)

The status of the newly created domain.

createDomainResponse_httpStatus :: Lens' CreateDomainResponse Int

The response's http status code.