Amazonka.OpenSearch.CreateOutboundConnection

Description

Creates a new cross-cluster connection from a local OpenSearch domain to a remote OpenSearch domain.

Creating a Request

data CreateOutboundConnection

Container for the parameters to the CreateOutboundConnection operation.

See: newCreateOutboundConnection smart constructor.

Constructors

CreateOutboundConnection'

Fields

localDomainInfo :: DomainInformationContainer

The AWSDomainInformation for the local OpenSearch domain.

remoteDomainInfo :: DomainInformationContainer

The AWSDomainInformation for the remote OpenSearch domain.

connectionAlias :: Text

The connection alias used used by the customer for this cross-cluster connection.

newCreateOutboundConnection

Arguments

:: DomainInformationContainer
$sel:localDomainInfo:CreateOutboundConnection'

-> DomainInformationContainer
$sel:remoteDomainInfo:CreateOutboundConnection'

-> Text
$sel:connectionAlias:CreateOutboundConnection'

-> CreateOutboundConnection

Create a value of CreateOutboundConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:localDomainInfo:CreateOutboundConnection', createOutboundConnection_localDomainInfo - The AWSDomainInformation for the local OpenSearch domain.

$sel:remoteDomainInfo:CreateOutboundConnection', createOutboundConnection_remoteDomainInfo - The AWSDomainInformation for the remote OpenSearch domain.

$sel:connectionAlias:CreateOutboundConnection', createOutboundConnection_connectionAlias - The connection alias used used by the customer for this cross-cluster connection.

Request Lenses

createOutboundConnection_localDomainInfo :: Lens' CreateOutboundConnection DomainInformationContainer

The AWSDomainInformation for the local OpenSearch domain.

createOutboundConnection_remoteDomainInfo :: Lens' CreateOutboundConnection DomainInformationContainer

The AWSDomainInformation for the remote OpenSearch domain.

createOutboundConnection_connectionAlias :: Lens' CreateOutboundConnection Text

The connection alias used used by the customer for this cross-cluster connection.

Destructuring the Response

data CreateOutboundConnectionResponse

The result of a CreateOutboundConnection request. Contains the details about the newly created cross-cluster connection.

See: newCreateOutboundConnectionResponse smart constructor.

Constructors

CreateOutboundConnectionResponse'

Fields

remoteDomainInfo :: Maybe DomainInformationContainer

The AWSDomainInformation for the remote OpenSearch domain.

connectionAlias :: Maybe Text

The connection alias provided during the create connection request.

localDomainInfo :: Maybe DomainInformationContainer

The AWSDomainInformation for the local OpenSearch domain.

connectionId :: Maybe Text

The unique ID for the created outbound connection, which is used for subsequent operations on the connection.

connectionStatus :: Maybe OutboundConnectionStatus

The OutboundConnectionStatus for the newly created connection.

httpStatus :: Int

The response's http status code.

newCreateOutboundConnectionResponse

Arguments

:: Int
$sel:httpStatus:CreateOutboundConnectionResponse'

-> CreateOutboundConnectionResponse

Create a value of CreateOutboundConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteDomainInfo:CreateOutboundConnection', createOutboundConnectionResponse_remoteDomainInfo - The AWSDomainInformation for the remote OpenSearch domain.

$sel:connectionAlias:CreateOutboundConnection', createOutboundConnectionResponse_connectionAlias - The connection alias provided during the create connection request.

$sel:localDomainInfo:CreateOutboundConnection', createOutboundConnectionResponse_localDomainInfo - The AWSDomainInformation for the local OpenSearch domain.

$sel:connectionId:CreateOutboundConnectionResponse', createOutboundConnectionResponse_connectionId - The unique ID for the created outbound connection, which is used for subsequent operations on the connection.

$sel:connectionStatus:CreateOutboundConnectionResponse', createOutboundConnectionResponse_connectionStatus - The OutboundConnectionStatus for the newly created connection.

$sel:httpStatus:CreateOutboundConnectionResponse', createOutboundConnectionResponse_httpStatus - The response's http status code.

Response Lenses

createOutboundConnectionResponse_remoteDomainInfo :: Lens' CreateOutboundConnectionResponse (Maybe DomainInformationContainer)

The AWSDomainInformation for the remote OpenSearch domain.

createOutboundConnectionResponse_connectionAlias :: Lens' CreateOutboundConnectionResponse (Maybe Text)

The connection alias provided during the create connection request.

createOutboundConnectionResponse_localDomainInfo :: Lens' CreateOutboundConnectionResponse (Maybe DomainInformationContainer)

The AWSDomainInformation for the local OpenSearch domain.

createOutboundConnectionResponse_connectionId :: Lens' CreateOutboundConnectionResponse (Maybe Text)

The unique ID for the created outbound connection, which is used for subsequent operations on the connection.

createOutboundConnectionResponse_connectionStatus :: Lens' CreateOutboundConnectionResponse (Maybe OutboundConnectionStatus)

The OutboundConnectionStatus for the newly created connection.

createOutboundConnectionResponse_httpStatus :: Lens' CreateOutboundConnectionResponse Int

The response's http status code.