Amazonka.OpenSearch.DescribeOutboundConnections

Description

Lists all the outbound cross-cluster connections for a local domain.

Creating a Request

data DescribeOutboundConnections

Container for the parameters to the DescribeOutboundConnections operation.

See: newDescribeOutboundConnections smart constructor.

Constructors

DescribeOutboundConnections'

Fields

filters :: Maybe [Filter]

A list of filters used to match properties for outbound cross-cluster connections. Available Filter names for this operation are:

connection-id

remote-domain-info.domain-name

remote-domain-info.owner-id

remote-domain-info.region

local-domain-info.domain-name

nextToken :: Maybe Text

NextToken is sent in case the earlier API call results contain the NextToken parameter. Used for pagination.

maxResults :: Maybe Int

Set this value to limit the number of results returned. If not specified, defaults to 100.

newDescribeOutboundConnections :: DescribeOutboundConnections

Create a value of DescribeOutboundConnections with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeOutboundConnections', describeOutboundConnections_filters - A list of filters used to match properties for outbound cross-cluster connections. Available Filter names for this operation are:

connection-id

remote-domain-info.domain-name

remote-domain-info.owner-id

remote-domain-info.region

local-domain-info.domain-name

$sel:nextToken:DescribeOutboundConnections', describeOutboundConnections_nextToken - NextToken is sent in case the earlier API call results contain the NextToken parameter. Used for pagination.

$sel:maxResults:DescribeOutboundConnections', describeOutboundConnections_maxResults - Set this value to limit the number of results returned. If not specified, defaults to 100.

Request Lenses

describeOutboundConnections_filters :: Lens' DescribeOutboundConnections (Maybe [Filter])

A list of filters used to match properties for outbound cross-cluster connections. Available Filter names for this operation are:

connection-id

remote-domain-info.domain-name

remote-domain-info.owner-id

remote-domain-info.region

local-domain-info.domain-name

describeOutboundConnections_nextToken :: Lens' DescribeOutboundConnections (Maybe Text)

NextToken is sent in case the earlier API call results contain the NextToken parameter. Used for pagination.

describeOutboundConnections_maxResults :: Lens' DescribeOutboundConnections (Maybe Int)

Set this value to limit the number of results returned. If not specified, defaults to 100.

Destructuring the Response

data DescribeOutboundConnectionsResponse

The result of a DescribeOutboundConnections request. Contains the list of connections matching the filter criteria.

See: newDescribeOutboundConnectionsResponse smart constructor.

Constructors

DescribeOutboundConnectionsResponse'

Fields

connections :: Maybe [OutboundConnection]

A list of OutboundConnection matching the specified filter criteria.

nextToken :: Maybe Text

If more results are available and NextToken is present, make the next request to the same API with the received NextToken to paginate the remaining results.

httpStatus :: Int

The response's http status code.

newDescribeOutboundConnectionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeOutboundConnectionsResponse'

-> DescribeOutboundConnectionsResponse

Create a value of DescribeOutboundConnectionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connections:DescribeOutboundConnectionsResponse', describeOutboundConnectionsResponse_connections - A list of OutboundConnection matching the specified filter criteria.

$sel:nextToken:DescribeOutboundConnections', describeOutboundConnectionsResponse_nextToken - If more results are available and NextToken is present, make the next request to the same API with the received NextToken to paginate the remaining results.

$sel:httpStatus:DescribeOutboundConnectionsResponse', describeOutboundConnectionsResponse_httpStatus - The response's http status code.

Response Lenses

describeOutboundConnectionsResponse_connections :: Lens' DescribeOutboundConnectionsResponse (Maybe [OutboundConnection])

A list of OutboundConnection matching the specified filter criteria.

describeOutboundConnectionsResponse_nextToken :: Lens' DescribeOutboundConnectionsResponse (Maybe Text)

If more results are available and NextToken is present, make the next request to the same API with the received NextToken to paginate the remaining results.

describeOutboundConnectionsResponse_httpStatus :: Lens' DescribeOutboundConnectionsResponse Int

The response's http status code.