Amazonka.OpenSearch.GetUpgradeHistory

Description

Retrieves the complete history of the last 10 upgrades performed on the domain.

Creating a Request

data GetUpgradeHistory

Container for the request parameters to the GetUpgradeHistory operation.

See: newGetUpgradeHistory smart constructor.

Constructors

GetUpgradeHistory'

Fields

nextToken :: Maybe Text

maxResults :: Maybe Int

domainName :: Text

newGetUpgradeHistory

Arguments

:: Text
$sel:domainName:GetUpgradeHistory'

-> GetUpgradeHistory

Create a value of GetUpgradeHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetUpgradeHistory', getUpgradeHistory_nextToken - Undocumented member.

$sel:maxResults:GetUpgradeHistory', getUpgradeHistory_maxResults - Undocumented member.

$sel:domainName:GetUpgradeHistory', getUpgradeHistory_domainName - Undocumented member.

Request Lenses

getUpgradeHistory_nextToken :: Lens' GetUpgradeHistory (Maybe Text)

Undocumented member.

getUpgradeHistory_maxResults :: Lens' GetUpgradeHistory (Maybe Int)

Undocumented member.

getUpgradeHistory_domainName :: Lens' GetUpgradeHistory Text

Undocumented member.

Destructuring the Response

data GetUpgradeHistoryResponse

Container for the response returned by the GetUpgradeHistory operation.

See: newGetUpgradeHistoryResponse smart constructor.

Constructors

GetUpgradeHistoryResponse'

Fields

nextToken :: Maybe Text

Pagination token that needs to be supplied to the next call to get the next page of results.

upgradeHistories :: Maybe [UpgradeHistory]

A list of UpgradeHistory objects corresponding to each upgrade or upgrade eligibility check performed on a domain returned as part of the GetUpgradeHistoryResponse object.

httpStatus :: Int

The response's http status code.

newGetUpgradeHistoryResponse

Arguments

:: Int
$sel:httpStatus:GetUpgradeHistoryResponse'

-> GetUpgradeHistoryResponse

Create a value of GetUpgradeHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetUpgradeHistory', getUpgradeHistoryResponse_nextToken - Pagination token that needs to be supplied to the next call to get the next page of results.

$sel:upgradeHistories:GetUpgradeHistoryResponse', getUpgradeHistoryResponse_upgradeHistories - A list of UpgradeHistory objects corresponding to each upgrade or upgrade eligibility check performed on a domain returned as part of the GetUpgradeHistoryResponse object.

$sel:httpStatus:GetUpgradeHistoryResponse', getUpgradeHistoryResponse_httpStatus - The response's http status code.

Response Lenses

getUpgradeHistoryResponse_nextToken :: Lens' GetUpgradeHistoryResponse (Maybe Text)

Pagination token that needs to be supplied to the next call to get the next page of results.

getUpgradeHistoryResponse_upgradeHistories :: Lens' GetUpgradeHistoryResponse (Maybe [UpgradeHistory])

A list of UpgradeHistory objects corresponding to each upgrade or upgrade eligibility check performed on a domain returned as part of the GetUpgradeHistoryResponse object.

getUpgradeHistoryResponse_httpStatus :: Lens' GetUpgradeHistoryResponse Int

The response's http status code.