Amazonka.OpenSearch.ListPackagesForDomain

Description

Lists all packages associated with the Amazon OpenSearch Service domain.

Creating a Request

data ListPackagesForDomain

Container for the request parameters to the ListPackagesForDomain operation.

See: newListPackagesForDomain smart constructor.

Constructors

ListPackagesForDomain'

Fields

nextToken :: Maybe Text

Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

maxResults :: Maybe Int

Limits results to a maximum number of packages.

domainName :: Text

The name of the domain for which you want to list associated packages.

newListPackagesForDomain

Arguments

:: Text
$sel:domainName:ListPackagesForDomain'

-> ListPackagesForDomain

Create a value of ListPackagesForDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPackagesForDomain', listPackagesForDomain_nextToken - Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

$sel:maxResults:ListPackagesForDomain', listPackagesForDomain_maxResults - Limits results to a maximum number of packages.

$sel:domainName:ListPackagesForDomain', listPackagesForDomain_domainName - The name of the domain for which you want to list associated packages.

Request Lenses

listPackagesForDomain_nextToken :: Lens' ListPackagesForDomain (Maybe Text)

Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

listPackagesForDomain_maxResults :: Lens' ListPackagesForDomain (Maybe Int)

Limits results to a maximum number of packages.

listPackagesForDomain_domainName :: Lens' ListPackagesForDomain Text

The name of the domain for which you want to list associated packages.

Destructuring the Response

data ListPackagesForDomainResponse

Container for the response parameters to the ListPackagesForDomain operation.

See: newListPackagesForDomainResponse smart constructor.

Constructors

ListPackagesForDomainResponse'

Fields

domainPackageDetailsList :: Maybe [DomainPackageDetails]

List of DomainPackageDetails objects.

nextToken :: Maybe Text

Pagination token to supply to the next call to get the next page of results.

httpStatus :: Int

The response's http status code.

newListPackagesForDomainResponse

Arguments

:: Int
$sel:httpStatus:ListPackagesForDomainResponse'

-> ListPackagesForDomainResponse

Create a value of ListPackagesForDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainPackageDetailsList:ListPackagesForDomainResponse', listPackagesForDomainResponse_domainPackageDetailsList - List of DomainPackageDetails objects.

$sel:nextToken:ListPackagesForDomain', listPackagesForDomainResponse_nextToken - Pagination token to supply to the next call to get the next page of results.

$sel:httpStatus:ListPackagesForDomainResponse', listPackagesForDomainResponse_httpStatus - The response's http status code.

Response Lenses

listPackagesForDomainResponse_domainPackageDetailsList :: Lens' ListPackagesForDomainResponse (Maybe [DomainPackageDetails])

List of DomainPackageDetails objects.

listPackagesForDomainResponse_nextToken :: Lens' ListPackagesForDomainResponse (Maybe Text)

Pagination token to supply to the next call to get the next page of results.

listPackagesForDomainResponse_httpStatus :: Lens' ListPackagesForDomainResponse Int

The response's http status code.