Amazonka.OpenSearch.PurchaseReservedInstanceOffering

Description

Allows you to purchase reserved OpenSearch instances.

Creating a Request

data PurchaseReservedInstanceOffering

Container for parameters to PurchaseReservedInstanceOffering

See: newPurchaseReservedInstanceOffering smart constructor.

Constructors

PurchaseReservedInstanceOffering'

Fields

instanceCount :: Maybe Natural

The number of OpenSearch instances to reserve.

reservedInstanceOfferingId :: Text

The ID of the reserved OpenSearch instance offering to purchase.

reservationName :: Text

A customer-specified identifier to track this reservation.

newPurchaseReservedInstanceOffering

Arguments

:: Text
$sel:reservedInstanceOfferingId:PurchaseReservedInstanceOffering'

-> Text
$sel:reservationName:PurchaseReservedInstanceOffering'

-> PurchaseReservedInstanceOffering

Create a value of PurchaseReservedInstanceOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceCount:PurchaseReservedInstanceOffering', purchaseReservedInstanceOffering_instanceCount - The number of OpenSearch instances to reserve.

$sel:reservedInstanceOfferingId:PurchaseReservedInstanceOffering', purchaseReservedInstanceOffering_reservedInstanceOfferingId - The ID of the reserved OpenSearch instance offering to purchase.

$sel:reservationName:PurchaseReservedInstanceOffering', purchaseReservedInstanceOffering_reservationName - A customer-specified identifier to track this reservation.

Request Lenses

purchaseReservedInstanceOffering_instanceCount :: Lens' PurchaseReservedInstanceOffering (Maybe Natural)

The number of OpenSearch instances to reserve.

purchaseReservedInstanceOffering_reservedInstanceOfferingId :: Lens' PurchaseReservedInstanceOffering Text

The ID of the reserved OpenSearch instance offering to purchase.

purchaseReservedInstanceOffering_reservationName :: Lens' PurchaseReservedInstanceOffering Text

A customer-specified identifier to track this reservation.

Destructuring the Response

data PurchaseReservedInstanceOfferingResponse

Represents the output of a PurchaseReservedInstanceOffering operation.

See: newPurchaseReservedInstanceOfferingResponse smart constructor.

Constructors

PurchaseReservedInstanceOfferingResponse'

Fields

reservationName :: Maybe Text

The customer-specified identifier used to track this reservation.

reservedInstanceId :: Maybe Text

Details of the reserved OpenSearch instance which was purchased.

httpStatus :: Int

The response's http status code.

newPurchaseReservedInstanceOfferingResponse

Arguments

:: Int
$sel:httpStatus:PurchaseReservedInstanceOfferingResponse'

-> PurchaseReservedInstanceOfferingResponse

Create a value of PurchaseReservedInstanceOfferingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservationName:PurchaseReservedInstanceOffering', purchaseReservedInstanceOfferingResponse_reservationName - The customer-specified identifier used to track this reservation.

$sel:reservedInstanceId:PurchaseReservedInstanceOfferingResponse', purchaseReservedInstanceOfferingResponse_reservedInstanceId - Details of the reserved OpenSearch instance which was purchased.

$sel:httpStatus:PurchaseReservedInstanceOfferingResponse', purchaseReservedInstanceOfferingResponse_httpStatus - The response's http status code.

Response Lenses

purchaseReservedInstanceOfferingResponse_reservationName :: Lens' PurchaseReservedInstanceOfferingResponse (Maybe Text)

The customer-specified identifier used to track this reservation.

purchaseReservedInstanceOfferingResponse_reservedInstanceId :: Lens' PurchaseReservedInstanceOfferingResponse (Maybe Text)

Details of the reserved OpenSearch instance which was purchased.

purchaseReservedInstanceOfferingResponse_httpStatus :: Lens' PurchaseReservedInstanceOfferingResponse Int

The response's http status code.