Amazonka.OpenSearch.RejectInboundConnection

Description

Allows the remote domain owner to reject an inbound cross-cluster connection request.

Creating a Request

data RejectInboundConnection

Container for the parameters to the RejectInboundConnection operation.

See: newRejectInboundConnection smart constructor.

Constructors

RejectInboundConnection'

Fields

connectionId :: Text

The ID of the inbound connection to reject.

newRejectInboundConnection

Arguments

:: Text
$sel:connectionId:RejectInboundConnection'

-> RejectInboundConnection

Create a value of RejectInboundConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionId:RejectInboundConnection', rejectInboundConnection_connectionId - The ID of the inbound connection to reject.

Request Lenses

rejectInboundConnection_connectionId :: Lens' RejectInboundConnection Text

The ID of the inbound connection to reject.

Destructuring the Response

data RejectInboundConnectionResponse

The result of a RejectInboundConnection operation. Contains details about the rejected inbound connection.

See: newRejectInboundConnectionResponse smart constructor.

Constructors

RejectInboundConnectionResponse'

Fields

connection :: Maybe InboundConnection

The InboundConnection of the rejected inbound connection.

httpStatus :: Int

The response's http status code.

newRejectInboundConnectionResponse

Arguments

:: Int
$sel:httpStatus:RejectInboundConnectionResponse'

-> RejectInboundConnectionResponse

Create a value of RejectInboundConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connection:RejectInboundConnectionResponse', rejectInboundConnectionResponse_connection - The InboundConnection of the rejected inbound connection.

$sel:httpStatus:RejectInboundConnectionResponse', rejectInboundConnectionResponse_httpStatus - The response's http status code.

Response Lenses

rejectInboundConnectionResponse_connection :: Lens' RejectInboundConnectionResponse (Maybe InboundConnection)

The InboundConnection of the rejected inbound connection.

rejectInboundConnectionResponse_httpStatus :: Lens' RejectInboundConnectionResponse Int

The response's http status code.