Amazonka.OpenSearch.Types.ClusterConfig

Description

Documentation

data ClusterConfig

The configuration for the domain cluster, such as the type and number of instances.

See: newClusterConfig smart constructor.

Constructors

ClusterConfig'

Fields

dedicatedMasterCount :: Maybe Int

Total number of dedicated master nodes, active and on standby, for the cluster.

dedicatedMasterType :: Maybe OpenSearchPartitionInstanceType

The instance type for a dedicated master node.

dedicatedMasterEnabled :: Maybe Bool

A boolean value to indicate whether a dedicated master node is enabled. See Dedicated master nodes in Amazon OpenSearch Service for more information.

instanceCount :: Maybe Int

The number of instances in the specified domain cluster.

coldStorageOptions :: Maybe ColdStorageOptions

Specifies the ColdStorageOptions config for a Domain

zoneAwarenessEnabled :: Maybe Bool

A boolean value to indicate whether zone awareness is enabled. See Configuring a multi-AZ domain in Amazon OpenSearch Service for more information.

instanceType :: Maybe OpenSearchPartitionInstanceType

The instance type for an OpenSearch cluster. UltraWarm instance types are not supported for data instances.

warmEnabled :: Maybe Bool

True to enable UltraWarm storage.

zoneAwarenessConfig :: Maybe ZoneAwarenessConfig

The zone awareness configuration for a domain when zone awareness is enabled.

warmCount :: Maybe Int

The number of UltraWarm nodes in the cluster.

warmType :: Maybe OpenSearchWarmPartitionInstanceType

The instance type for the OpenSearch cluster's warm nodes.

newClusterConfig :: ClusterConfig

Create a value of ClusterConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dedicatedMasterCount:ClusterConfig', clusterConfig_dedicatedMasterCount - Total number of dedicated master nodes, active and on standby, for the cluster.

$sel:dedicatedMasterType:ClusterConfig', clusterConfig_dedicatedMasterType - The instance type for a dedicated master node.

$sel:dedicatedMasterEnabled:ClusterConfig', clusterConfig_dedicatedMasterEnabled - A boolean value to indicate whether a dedicated master node is enabled. See Dedicated master nodes in Amazon OpenSearch Service for more information.

$sel:instanceCount:ClusterConfig', clusterConfig_instanceCount - The number of instances in the specified domain cluster.

$sel:coldStorageOptions:ClusterConfig', clusterConfig_coldStorageOptions - Specifies the ColdStorageOptions config for a Domain

$sel:zoneAwarenessEnabled:ClusterConfig', clusterConfig_zoneAwarenessEnabled - A boolean value to indicate whether zone awareness is enabled. See Configuring a multi-AZ domain in Amazon OpenSearch Service for more information.

$sel:instanceType:ClusterConfig', clusterConfig_instanceType - The instance type for an OpenSearch cluster. UltraWarm instance types are not supported for data instances.

$sel:warmEnabled:ClusterConfig', clusterConfig_warmEnabled - True to enable UltraWarm storage.

$sel:zoneAwarenessConfig:ClusterConfig', clusterConfig_zoneAwarenessConfig - The zone awareness configuration for a domain when zone awareness is enabled.

$sel:warmCount:ClusterConfig', clusterConfig_warmCount - The number of UltraWarm nodes in the cluster.

$sel:warmType:ClusterConfig', clusterConfig_warmType - The instance type for the OpenSearch cluster's warm nodes.

clusterConfig_dedicatedMasterCount :: Lens' ClusterConfig (Maybe Int)

Total number of dedicated master nodes, active and on standby, for the cluster.

clusterConfig_dedicatedMasterType :: Lens' ClusterConfig (Maybe OpenSearchPartitionInstanceType)

The instance type for a dedicated master node.

clusterConfig_dedicatedMasterEnabled :: Lens' ClusterConfig (Maybe Bool)

A boolean value to indicate whether a dedicated master node is enabled. See Dedicated master nodes in Amazon OpenSearch Service for more information.

clusterConfig_instanceCount :: Lens' ClusterConfig (Maybe Int)

The number of instances in the specified domain cluster.

clusterConfig_coldStorageOptions :: Lens' ClusterConfig (Maybe ColdStorageOptions)

Specifies the ColdStorageOptions config for a Domain

clusterConfig_zoneAwarenessEnabled :: Lens' ClusterConfig (Maybe Bool)

A boolean value to indicate whether zone awareness is enabled. See Configuring a multi-AZ domain in Amazon OpenSearch Service for more information.

clusterConfig_instanceType :: Lens' ClusterConfig (Maybe OpenSearchPartitionInstanceType)

The instance type for an OpenSearch cluster. UltraWarm instance types are not supported for data instances.

clusterConfig_warmEnabled :: Lens' ClusterConfig (Maybe Bool)

True to enable UltraWarm storage.

clusterConfig_zoneAwarenessConfig :: Lens' ClusterConfig (Maybe ZoneAwarenessConfig)

The zone awareness configuration for a domain when zone awareness is enabled.

clusterConfig_warmCount :: Lens' ClusterConfig (Maybe Int)

The number of UltraWarm nodes in the cluster.

clusterConfig_warmType :: Lens' ClusterConfig (Maybe OpenSearchWarmPartitionInstanceType)

The instance type for the OpenSearch cluster's warm nodes.