Amazonka.OpenSearch.Types.DomainPackageDetails

Description

Documentation

data DomainPackageDetails

Information on a package associated with a domain.

See: newDomainPackageDetails smart constructor.

Constructors

DomainPackageDetails'

Fields

lastUpdated :: Maybe POSIX

The timestamp of the most recent update to the package association status.

packageID :: Maybe Text

The internal ID of the package.

packageType :: Maybe PackageType

Currently supports only TXT-DICTIONARY.

packageName :: Maybe Text

User-specified name of the package.

packageVersion :: Maybe Text

domainPackageStatus :: Maybe DomainPackageStatus

State of the association. Values are ASSOCIATING, ASSOCIATION_FAILED, ACTIVE, DISSOCIATING, and DISSOCIATION_FAILED.

domainName :: Maybe Text

The name of the domain you've associated a package with.

errorDetails :: Maybe ErrorDetails

Additional information if the package is in an error state. Null otherwise.

referencePath :: Maybe Text

The relative path on Amazon OpenSearch Service nodes, which can be used as synonym_path when the package is a synonym file.

newDomainPackageDetails :: DomainPackageDetails

Create a value of DomainPackageDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdated:DomainPackageDetails', domainPackageDetails_lastUpdated - The timestamp of the most recent update to the package association status.

$sel:packageID:DomainPackageDetails', domainPackageDetails_packageID - The internal ID of the package.

$sel:packageType:DomainPackageDetails', domainPackageDetails_packageType - Currently supports only TXT-DICTIONARY.

$sel:packageName:DomainPackageDetails', domainPackageDetails_packageName - User-specified name of the package.

$sel:packageVersion:DomainPackageDetails', domainPackageDetails_packageVersion - Undocumented member.

$sel:domainPackageStatus:DomainPackageDetails', domainPackageDetails_domainPackageStatus - State of the association. Values are ASSOCIATING, ASSOCIATION_FAILED, ACTIVE, DISSOCIATING, and DISSOCIATION_FAILED.

$sel:domainName:DomainPackageDetails', domainPackageDetails_domainName - The name of the domain you've associated a package with.

$sel:errorDetails:DomainPackageDetails', domainPackageDetails_errorDetails - Additional information if the package is in an error state. Null otherwise.

$sel:referencePath:DomainPackageDetails', domainPackageDetails_referencePath - The relative path on Amazon OpenSearch Service nodes, which can be used as synonym_path when the package is a synonym file.

domainPackageDetails_lastUpdated :: Lens' DomainPackageDetails (Maybe UTCTime)

The timestamp of the most recent update to the package association status.

domainPackageDetails_packageID :: Lens' DomainPackageDetails (Maybe Text)

The internal ID of the package.

domainPackageDetails_packageType :: Lens' DomainPackageDetails (Maybe PackageType)

Currently supports only TXT-DICTIONARY.

domainPackageDetails_packageName :: Lens' DomainPackageDetails (Maybe Text)

User-specified name of the package.

domainPackageDetails_packageVersion :: Lens' DomainPackageDetails (Maybe Text)

Undocumented member.

domainPackageDetails_domainPackageStatus :: Lens' DomainPackageDetails (Maybe DomainPackageStatus)

State of the association. Values are ASSOCIATING, ASSOCIATION_FAILED, ACTIVE, DISSOCIATING, and DISSOCIATION_FAILED.

domainPackageDetails_domainName :: Lens' DomainPackageDetails (Maybe Text)

The name of the domain you've associated a package with.

domainPackageDetails_errorDetails :: Lens' DomainPackageDetails (Maybe ErrorDetails)

Additional information if the package is in an error state. Null otherwise.

domainPackageDetails_referencePath :: Lens' DomainPackageDetails (Maybe Text)

The relative path on Amazon OpenSearch Service nodes, which can be used as synonym_path when the package is a synonym file.