Amazonka.OpenSearch.Types.DomainStatus

Description

Documentation

data DomainStatus

The current status of a domain.

See: newDomainStatus smart constructor.

Constructors

DomainStatus'

Fields

eBSOptions :: Maybe EBSOptions

The EBSOptions for the specified domain.

engineVersion :: Maybe Text

nodeToNodeEncryptionOptions :: Maybe NodeToNodeEncryptionOptions

The status of the NodeToNodeEncryptionOptions.

accessPolicies :: Maybe Text

IAM access policy as a JSON-formatted string.

serviceSoftwareOptions :: Maybe ServiceSoftwareOptions

The current status of the domain's service software.

autoTuneOptions :: Maybe AutoTuneOptionsOutput

The current status of the domain's Auto-Tune options.

logPublishingOptions :: Maybe (HashMap LogType LogPublishingOption)

Log publishing options for the given domain.

advancedSecurityOptions :: Maybe AdvancedSecurityOptions

The current status of the domain's advanced security options.

created :: Maybe Bool

The domain creation status. True if the creation of a domain is complete. False if domain creation is still in progress.

snapshotOptions :: Maybe SnapshotOptions

The status of the SnapshotOptions.

cognitoOptions :: Maybe CognitoOptions

The CognitoOptions for the specified domain. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

encryptionAtRestOptions :: Maybe EncryptionAtRestOptions

The status of the EncryptionAtRestOptions.

deleted :: Maybe Bool

The domain deletion status. True if a delete request has been received for the domain but resource cleanup is still in progress. False if the domain has not been deleted. Once domain deletion is complete, the status of the domain is no longer returned.

vPCOptions :: Maybe VPCDerivedInfo

The VPCOptions for the specified domain. For more information, see Launching your Amazon OpenSearch Service domains using a VPC.

endpoints :: Maybe (HashMap Text Text)

Map containing the domain endpoints used to submit index and search requests. Example key, value: 'vpc','vpc-endpoint-h2dsd34efgyghrtguk5gt6j2foh4.us-east-1.es.amazonaws.com'.

domainEndpointOptions :: Maybe DomainEndpointOptions

The current status of the domain's endpoint options.

processing :: Maybe Bool

The status of the domain configuration. True if Amazon OpenSearch Service is processing configuration changes. False if the configuration is active.

endpoint :: Maybe Text

The domain endpoint that you use to submit index and search requests.

upgradeProcessing :: Maybe Bool

The status of a domain version upgrade. True if Amazon OpenSearch Service is undergoing a version upgrade. False if the configuration is active.

advancedOptions :: Maybe (HashMap Text Text)

The status of the AdvancedOptions.

domainId :: Text

The unique identifier for the specified domain.

domainName :: Text

The name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

arn :: Text

The Amazon Resource Name (ARN) of a domain. See IAM identifiers in the AWS Identity and Access Management User Guide for more information.

clusterConfig :: ClusterConfig

The type and number of instances in the domain.

newDomainStatus

Arguments

:: Text
$sel:domainId:DomainStatus'

-> Text
$sel:domainName:DomainStatus'

-> Text
$sel:arn:DomainStatus'

-> ClusterConfig
$sel:clusterConfig:DomainStatus'

-> DomainStatus

Create a value of DomainStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eBSOptions:DomainStatus', domainStatus_eBSOptions - The EBSOptions for the specified domain.

$sel:engineVersion:DomainStatus', domainStatus_engineVersion - Undocumented member.

$sel:nodeToNodeEncryptionOptions:DomainStatus', domainStatus_nodeToNodeEncryptionOptions - The status of the NodeToNodeEncryptionOptions.

$sel:accessPolicies:DomainStatus', domainStatus_accessPolicies - IAM access policy as a JSON-formatted string.

$sel:serviceSoftwareOptions:DomainStatus', domainStatus_serviceSoftwareOptions - The current status of the domain's service software.

$sel:autoTuneOptions:DomainStatus', domainStatus_autoTuneOptions - The current status of the domain's Auto-Tune options.

$sel:logPublishingOptions:DomainStatus', domainStatus_logPublishingOptions - Log publishing options for the given domain.

$sel:advancedSecurityOptions:DomainStatus', domainStatus_advancedSecurityOptions - The current status of the domain's advanced security options.

$sel:created:DomainStatus', domainStatus_created - The domain creation status. True if the creation of a domain is complete. False if domain creation is still in progress.

$sel:snapshotOptions:DomainStatus', domainStatus_snapshotOptions - The status of the SnapshotOptions.

$sel:cognitoOptions:DomainStatus', domainStatus_cognitoOptions - The CognitoOptions for the specified domain. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

$sel:encryptionAtRestOptions:DomainStatus', domainStatus_encryptionAtRestOptions - The status of the EncryptionAtRestOptions.

$sel:deleted:DomainStatus', domainStatus_deleted - The domain deletion status. True if a delete request has been received for the domain but resource cleanup is still in progress. False if the domain has not been deleted. Once domain deletion is complete, the status of the domain is no longer returned.

$sel:vPCOptions:DomainStatus', domainStatus_vPCOptions - The VPCOptions for the specified domain. For more information, see Launching your Amazon OpenSearch Service domains using a VPC.

$sel:endpoints:DomainStatus', domainStatus_endpoints - Map containing the domain endpoints used to submit index and search requests. Example key, value: 'vpc','vpc-endpoint-h2dsd34efgyghrtguk5gt6j2foh4.us-east-1.es.amazonaws.com'.

$sel:domainEndpointOptions:DomainStatus', domainStatus_domainEndpointOptions - The current status of the domain's endpoint options.

$sel:processing:DomainStatus', domainStatus_processing - The status of the domain configuration. True if Amazon OpenSearch Service is processing configuration changes. False if the configuration is active.

$sel:endpoint:DomainStatus', domainStatus_endpoint - The domain endpoint that you use to submit index and search requests.

$sel:upgradeProcessing:DomainStatus', domainStatus_upgradeProcessing - The status of a domain version upgrade. True if Amazon OpenSearch Service is undergoing a version upgrade. False if the configuration is active.

$sel:advancedOptions:DomainStatus', domainStatus_advancedOptions - The status of the AdvancedOptions.

$sel:domainId:DomainStatus', domainStatus_domainId - The unique identifier for the specified domain.

$sel:domainName:DomainStatus', domainStatus_domainName - The name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

$sel:arn:DomainStatus', domainStatus_arn - The Amazon Resource Name (ARN) of a domain. See IAM identifiers in the AWS Identity and Access Management User Guide for more information.

$sel:clusterConfig:DomainStatus', domainStatus_clusterConfig - The type and number of instances in the domain.

domainStatus_eBSOptions :: Lens' DomainStatus (Maybe EBSOptions)

The EBSOptions for the specified domain.

domainStatus_engineVersion :: Lens' DomainStatus (Maybe Text)

Undocumented member.

domainStatus_nodeToNodeEncryptionOptions :: Lens' DomainStatus (Maybe NodeToNodeEncryptionOptions)

The status of the NodeToNodeEncryptionOptions.

domainStatus_accessPolicies :: Lens' DomainStatus (Maybe Text)

IAM access policy as a JSON-formatted string.

domainStatus_serviceSoftwareOptions :: Lens' DomainStatus (Maybe ServiceSoftwareOptions)

The current status of the domain's service software.

domainStatus_autoTuneOptions :: Lens' DomainStatus (Maybe AutoTuneOptionsOutput)

The current status of the domain's Auto-Tune options.

domainStatus_logPublishingOptions :: Lens' DomainStatus (Maybe (HashMap LogType LogPublishingOption))

Log publishing options for the given domain.

domainStatus_advancedSecurityOptions :: Lens' DomainStatus (Maybe AdvancedSecurityOptions)

The current status of the domain's advanced security options.

domainStatus_created :: Lens' DomainStatus (Maybe Bool)

The domain creation status. True if the creation of a domain is complete. False if domain creation is still in progress.

domainStatus_snapshotOptions :: Lens' DomainStatus (Maybe SnapshotOptions)

The status of the SnapshotOptions.

domainStatus_cognitoOptions :: Lens' DomainStatus (Maybe CognitoOptions)

The CognitoOptions for the specified domain. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

domainStatus_encryptionAtRestOptions :: Lens' DomainStatus (Maybe EncryptionAtRestOptions)

The status of the EncryptionAtRestOptions.

domainStatus_deleted :: Lens' DomainStatus (Maybe Bool)

The domain deletion status. True if a delete request has been received for the domain but resource cleanup is still in progress. False if the domain has not been deleted. Once domain deletion is complete, the status of the domain is no longer returned.

domainStatus_vPCOptions :: Lens' DomainStatus (Maybe VPCDerivedInfo)

The VPCOptions for the specified domain. For more information, see Launching your Amazon OpenSearch Service domains using a VPC.

domainStatus_endpoints :: Lens' DomainStatus (Maybe (HashMap Text Text))

Map containing the domain endpoints used to submit index and search requests. Example key, value: 'vpc','vpc-endpoint-h2dsd34efgyghrtguk5gt6j2foh4.us-east-1.es.amazonaws.com'.

domainStatus_domainEndpointOptions :: Lens' DomainStatus (Maybe DomainEndpointOptions)

The current status of the domain's endpoint options.

domainStatus_processing :: Lens' DomainStatus (Maybe Bool)

The status of the domain configuration. True if Amazon OpenSearch Service is processing configuration changes. False if the configuration is active.

domainStatus_endpoint :: Lens' DomainStatus (Maybe Text)

The domain endpoint that you use to submit index and search requests.

domainStatus_upgradeProcessing :: Lens' DomainStatus (Maybe Bool)

The status of a domain version upgrade. True if Amazon OpenSearch Service is undergoing a version upgrade. False if the configuration is active.

domainStatus_advancedOptions :: Lens' DomainStatus (Maybe (HashMap Text Text))

The status of the AdvancedOptions.

domainStatus_domainId :: Lens' DomainStatus Text

The unique identifier for the specified domain.

domainStatus_domainName :: Lens' DomainStatus Text

The name of a domain. Domain names are unique across the domains owned by an account within an AWS region. Domain names start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

domainStatus_arn :: Lens' DomainStatus Text

The Amazon Resource Name (ARN) of a domain. See IAM identifiers in the AWS Identity and Access Management User Guide for more information.

domainStatus_clusterConfig :: Lens' DomainStatus ClusterConfig

The type and number of instances in the domain.