Amazonka.OpenSearch.Types.InboundConnectionStatus

Description

Documentation

data InboundConnectionStatus

The connection status of an inbound cross-cluster connection.

See: newInboundConnectionStatus smart constructor.

Constructors

InboundConnectionStatus'

Fields

message :: Maybe Text

Verbose information for the inbound connection status.

statusCode :: Maybe InboundConnectionStatusCode

The state code for the inbound connection. Can be one of the following:

PENDING_ACCEPTANCE: Inbound connection is not yet accepted by the remote domain owner.

APPROVED: Inbound connection is pending acceptance by the remote domain owner.

PROVISIONING: Inbound connection provisioning is in progress.

ACTIVE: Inbound connection is active and ready to use.

REJECTING: Inbound connection rejection is in process.

REJECTED: Inbound connection is rejected.

DELETING: Inbound connection deletion is in progress.

DELETED: Inbound connection is deleted and can no longer be used.

newInboundConnectionStatus :: InboundConnectionStatus

Create a value of InboundConnectionStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:InboundConnectionStatus', inboundConnectionStatus_message - Verbose information for the inbound connection status.

$sel:statusCode:InboundConnectionStatus', inboundConnectionStatus_statusCode - The state code for the inbound connection. Can be one of the following:

PENDING_ACCEPTANCE: Inbound connection is not yet accepted by the remote domain owner.

APPROVED: Inbound connection is pending acceptance by the remote domain owner.

PROVISIONING: Inbound connection provisioning is in progress.

ACTIVE: Inbound connection is active and ready to use.

REJECTING: Inbound connection rejection is in process.

REJECTED: Inbound connection is rejected.

DELETING: Inbound connection deletion is in progress.

DELETED: Inbound connection is deleted and can no longer be used.

inboundConnectionStatus_message :: Lens' InboundConnectionStatus (Maybe Text)

Verbose information for the inbound connection status.

inboundConnectionStatus_statusCode :: Lens' InboundConnectionStatus (Maybe InboundConnectionStatusCode)

The state code for the inbound connection. Can be one of the following:

PENDING_ACCEPTANCE: Inbound connection is not yet accepted by the remote domain owner.

APPROVED: Inbound connection is pending acceptance by the remote domain owner.

PROVISIONING: Inbound connection provisioning is in progress.

ACTIVE: Inbound connection is active and ready to use.

REJECTING: Inbound connection rejection is in process.

REJECTED: Inbound connection is rejected.

DELETING: Inbound connection deletion is in progress.

DELETED: Inbound connection is deleted and can no longer be used.