Amazonka.OpenSearch.Types.InstanceCountLimits

Description

Documentation

data InstanceCountLimits

InstanceCountLimits represents the limits on the number of instances that can be created in Amazon OpenSearch Service for a given InstanceType.

See: newInstanceCountLimits smart constructor.

Constructors

InstanceCountLimits'

Fields

maximumInstanceCount :: Maybe Int

minimumInstanceCount :: Maybe Int

newInstanceCountLimits :: InstanceCountLimits

Create a value of InstanceCountLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maximumInstanceCount:InstanceCountLimits', instanceCountLimits_maximumInstanceCount - Undocumented member.

$sel:minimumInstanceCount:InstanceCountLimits', instanceCountLimits_minimumInstanceCount - Undocumented member.

instanceCountLimits_maximumInstanceCount :: Lens' InstanceCountLimits (Maybe Int)

Undocumented member.

instanceCountLimits_minimumInstanceCount :: Lens' InstanceCountLimits (Maybe Int)

Undocumented member.