Amazonka.OpenSearch.Types.RecurringCharge

Description

Documentation

data RecurringCharge

Contains the specific price and frequency of a recurring charges for a reserved OpenSearch instance, or for a reserved OpenSearch instance offering.

See: newRecurringCharge smart constructor.

Constructors

RecurringCharge'

Fields

recurringChargeFrequency :: Maybe Text

The frequency of the recurring charge.

recurringChargeAmount :: Maybe Double

The monetary amount of the recurring charge.

newRecurringCharge :: RecurringCharge

Create a value of RecurringCharge with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recurringChargeFrequency:RecurringCharge', recurringCharge_recurringChargeFrequency - The frequency of the recurring charge.

$sel:recurringChargeAmount:RecurringCharge', recurringCharge_recurringChargeAmount - The monetary amount of the recurring charge.

recurringCharge_recurringChargeFrequency :: Lens' RecurringCharge (Maybe Text)

The frequency of the recurring charge.

recurringCharge_recurringChargeAmount :: Lens' RecurringCharge (Maybe Double)

The monetary amount of the recurring charge.