Amazonka.OpenSearch.Types.ReservedInstance

Description

Documentation

data ReservedInstance

Details of a reserved OpenSearch instance.

See: newReservedInstance smart constructor.

Constructors

ReservedInstance'

Fields

state :: Maybe Text

The state of the reserved OpenSearch instance.

currencyCode :: Maybe Text

The currency code for the reserved OpenSearch instance offering.

instanceCount :: Maybe Int

The number of OpenSearch instances that have been reserved.

startTime :: Maybe POSIX

The time the reservation started.

instanceType :: Maybe OpenSearchPartitionInstanceType

The OpenSearch instance type offered by the reserved instance offering.

reservationName :: Maybe Text

The customer-specified identifier to track this reservation.

billingSubscriptionId :: Maybe Integer

recurringCharges :: Maybe [RecurringCharge]

The charge to your account regardless of whether you are creating any domains using the instance offering.

usagePrice :: Maybe Double

The rate you are charged for each hour for the domain that is using this reserved instance.

reservedInstanceId :: Maybe Text

The unique identifier for the reservation.

reservedInstanceOfferingId :: Maybe Text

The offering identifier.

fixedPrice :: Maybe Double

The upfront fixed charge you will paid to purchase the specific reserved OpenSearch instance offering.

duration :: Maybe Int

The duration, in seconds, for which the OpenSearch instance is reserved.

paymentOption :: Maybe ReservedInstancePaymentOption

The payment option as defined in the reserved OpenSearch instance offering.

newReservedInstance :: ReservedInstance

Create a value of ReservedInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ReservedInstance', reservedInstance_state - The state of the reserved OpenSearch instance.

$sel:currencyCode:ReservedInstance', reservedInstance_currencyCode - The currency code for the reserved OpenSearch instance offering.

$sel:instanceCount:ReservedInstance', reservedInstance_instanceCount - The number of OpenSearch instances that have been reserved.

$sel:startTime:ReservedInstance', reservedInstance_startTime - The time the reservation started.

$sel:instanceType:ReservedInstance', reservedInstance_instanceType - The OpenSearch instance type offered by the reserved instance offering.

$sel:reservationName:ReservedInstance', reservedInstance_reservationName - The customer-specified identifier to track this reservation.

$sel:billingSubscriptionId:ReservedInstance', reservedInstance_billingSubscriptionId - Undocumented member.

$sel:recurringCharges:ReservedInstance', reservedInstance_recurringCharges - The charge to your account regardless of whether you are creating any domains using the instance offering.

$sel:usagePrice:ReservedInstance', reservedInstance_usagePrice - The rate you are charged for each hour for the domain that is using this reserved instance.

$sel:reservedInstanceId:ReservedInstance', reservedInstance_reservedInstanceId - The unique identifier for the reservation.

$sel:reservedInstanceOfferingId:ReservedInstance', reservedInstance_reservedInstanceOfferingId - The offering identifier.

$sel:fixedPrice:ReservedInstance', reservedInstance_fixedPrice - The upfront fixed charge you will paid to purchase the specific reserved OpenSearch instance offering.

$sel:duration:ReservedInstance', reservedInstance_duration - The duration, in seconds, for which the OpenSearch instance is reserved.

$sel:paymentOption:ReservedInstance', reservedInstance_paymentOption - The payment option as defined in the reserved OpenSearch instance offering.

reservedInstance_state :: Lens' ReservedInstance (Maybe Text)

The state of the reserved OpenSearch instance.

reservedInstance_currencyCode :: Lens' ReservedInstance (Maybe Text)

The currency code for the reserved OpenSearch instance offering.

reservedInstance_instanceCount :: Lens' ReservedInstance (Maybe Int)

The number of OpenSearch instances that have been reserved.

reservedInstance_startTime :: Lens' ReservedInstance (Maybe UTCTime)

The time the reservation started.

reservedInstance_instanceType :: Lens' ReservedInstance (Maybe OpenSearchPartitionInstanceType)

The OpenSearch instance type offered by the reserved instance offering.

reservedInstance_reservationName :: Lens' ReservedInstance (Maybe Text)

The customer-specified identifier to track this reservation.

reservedInstance_billingSubscriptionId :: Lens' ReservedInstance (Maybe Integer)

Undocumented member.

reservedInstance_recurringCharges :: Lens' ReservedInstance (Maybe [RecurringCharge])

The charge to your account regardless of whether you are creating any domains using the instance offering.

reservedInstance_usagePrice :: Lens' ReservedInstance (Maybe Double)

The rate you are charged for each hour for the domain that is using this reserved instance.

reservedInstance_reservedInstanceId :: Lens' ReservedInstance (Maybe Text)

The unique identifier for the reservation.

reservedInstance_reservedInstanceOfferingId :: Lens' ReservedInstance (Maybe Text)

The offering identifier.

reservedInstance_fixedPrice :: Lens' ReservedInstance (Maybe Double)

The upfront fixed charge you will paid to purchase the specific reserved OpenSearch instance offering.

reservedInstance_duration :: Lens' ReservedInstance (Maybe Int)

The duration, in seconds, for which the OpenSearch instance is reserved.

reservedInstance_paymentOption :: Lens' ReservedInstance (Maybe ReservedInstancePaymentOption)

The payment option as defined in the reserved OpenSearch instance offering.