Amazonka.OpenSearch.Types.SAMLOptionsOutput

Description

Documentation

data SAMLOptionsOutput

Describes the SAML application configured for the domain.

See: newSAMLOptionsOutput smart constructor.

Constructors

SAMLOptionsOutput'

Fields

enabled :: Maybe Bool

True if SAML is enabled.

idp :: Maybe SAMLIdp

Describes the SAML identity provider's information.

rolesKey :: Maybe Text

The key used for matching the SAML roles attribute.

sessionTimeoutMinutes :: Maybe Int

The duration, in minutes, after which a user session becomes inactive.

subjectKey :: Maybe Text

The key used for matching the SAML subject attribute.

newSAMLOptionsOutput :: SAMLOptionsOutput

Create a value of SAMLOptionsOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:SAMLOptionsOutput', sAMLOptionsOutput_enabled - True if SAML is enabled.

$sel:idp:SAMLOptionsOutput', sAMLOptionsOutput_idp - Describes the SAML identity provider's information.

$sel:rolesKey:SAMLOptionsOutput', sAMLOptionsOutput_rolesKey - The key used for matching the SAML roles attribute.

$sel:sessionTimeoutMinutes:SAMLOptionsOutput', sAMLOptionsOutput_sessionTimeoutMinutes - The duration, in minutes, after which a user session becomes inactive.

$sel:subjectKey:SAMLOptionsOutput', sAMLOptionsOutput_subjectKey - The key used for matching the SAML subject attribute.

sAMLOptionsOutput_enabled :: Lens' SAMLOptionsOutput (Maybe Bool)

True if SAML is enabled.

sAMLOptionsOutput_idp :: Lens' SAMLOptionsOutput (Maybe SAMLIdp)

Describes the SAML identity provider's information.

sAMLOptionsOutput_rolesKey :: Lens' SAMLOptionsOutput (Maybe Text)

The key used for matching the SAML roles attribute.

sAMLOptionsOutput_sessionTimeoutMinutes :: Lens' SAMLOptionsOutput (Maybe Int)

The duration, in minutes, after which a user session becomes inactive.

sAMLOptionsOutput_subjectKey :: Lens' SAMLOptionsOutput (Maybe Text)

The key used for matching the SAML subject attribute.