Amazonka.OpenSearch.Types.ServiceSoftwareOptions

Description

Documentation

data ServiceSoftwareOptions

The current options of an domain service software options.

See: newServiceSoftwareOptions smart constructor.

Constructors

ServiceSoftwareOptions'

Fields

automatedUpdateDate :: Maybe POSIX

The timestamp, in Epoch time, until which you can manually request a service software update. After this date, we automatically update your service software.

currentVersion :: Maybe Text

The current service software version present on the domain.

optionalDeployment :: Maybe Bool

True if a service software is never automatically updated. False if a service software is automatically updated after AutomatedUpdateDate.

updateStatus :: Maybe DeploymentStatus

The status of your service software update. This field can take the following values: ELIGIBLE, PENDING_UPDATE, IN_PROGRESS, COMPLETED, and NOT_ELIGIBLE.

cancellable :: Maybe Bool

True if you're able to cancel your service software version update. False if you can't cancel your service software update.

updateAvailable :: Maybe Bool

True if you're able to update your service software version. False if you can't update your service software version.

description :: Maybe Text

The description of the UpdateStatus.

newVersion' :: Maybe Text

The new service software version if one is available.

newServiceSoftwareOptions :: ServiceSoftwareOptions

Create a value of ServiceSoftwareOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:automatedUpdateDate:ServiceSoftwareOptions', serviceSoftwareOptions_automatedUpdateDate - The timestamp, in Epoch time, until which you can manually request a service software update. After this date, we automatically update your service software.

$sel:currentVersion:ServiceSoftwareOptions', serviceSoftwareOptions_currentVersion - The current service software version present on the domain.

$sel:optionalDeployment:ServiceSoftwareOptions', serviceSoftwareOptions_optionalDeployment - True if a service software is never automatically updated. False if a service software is automatically updated after AutomatedUpdateDate.

$sel:updateStatus:ServiceSoftwareOptions', serviceSoftwareOptions_updateStatus - The status of your service software update. This field can take the following values: ELIGIBLE, PENDING_UPDATE, IN_PROGRESS, COMPLETED, and NOT_ELIGIBLE.

$sel:cancellable:ServiceSoftwareOptions', serviceSoftwareOptions_cancellable - True if you're able to cancel your service software version update. False if you can't cancel your service software update.

$sel:updateAvailable:ServiceSoftwareOptions', serviceSoftwareOptions_updateAvailable - True if you're able to update your service software version. False if you can't update your service software version.

$sel:description:ServiceSoftwareOptions', serviceSoftwareOptions_description - The description of the UpdateStatus.

$sel:newVersion':ServiceSoftwareOptions', serviceSoftwareOptions_newVersion - The new service software version if one is available.

serviceSoftwareOptions_automatedUpdateDate :: Lens' ServiceSoftwareOptions (Maybe UTCTime)

The timestamp, in Epoch time, until which you can manually request a service software update. After this date, we automatically update your service software.

serviceSoftwareOptions_currentVersion :: Lens' ServiceSoftwareOptions (Maybe Text)

The current service software version present on the domain.

serviceSoftwareOptions_optionalDeployment :: Lens' ServiceSoftwareOptions (Maybe Bool)

True if a service software is never automatically updated. False if a service software is automatically updated after AutomatedUpdateDate.

serviceSoftwareOptions_updateStatus :: Lens' ServiceSoftwareOptions (Maybe DeploymentStatus)

The status of your service software update. This field can take the following values: ELIGIBLE, PENDING_UPDATE, IN_PROGRESS, COMPLETED, and NOT_ELIGIBLE.

serviceSoftwareOptions_cancellable :: Lens' ServiceSoftwareOptions (Maybe Bool)

True if you're able to cancel your service software version update. False if you can't cancel your service software update.

serviceSoftwareOptions_updateAvailable :: Lens' ServiceSoftwareOptions (Maybe Bool)

True if you're able to update your service software version. False if you can't update your service software version.

serviceSoftwareOptions_description :: Lens' ServiceSoftwareOptions (Maybe Text)

The description of the UpdateStatus.

serviceSoftwareOptions_newVersion :: Lens' ServiceSoftwareOptions (Maybe Text)

The new service software version if one is available.