Amazonka.OpenSearch.UpdateDomainConfig

Description

Modifies the cluster configuration of the specified domain, such as setting the instance type and the number of instances.

Creating a Request

data UpdateDomainConfig

Container for the parameters to the UpdateDomain operation. Specifies the type and number of instances in the domain cluster.

See: newUpdateDomainConfig smart constructor.

Constructors

UpdateDomainConfig'

Fields

eBSOptions :: Maybe EBSOptions

Specify the type and size of the EBS volume to use.

nodeToNodeEncryptionOptions :: Maybe NodeToNodeEncryptionOptions

Specifies node-to-node encryption options.

accessPolicies :: Maybe Text

IAM access policy as a JSON-formatted string.

autoTuneOptions :: Maybe AutoTuneOptions

Specifies Auto-Tune options.

logPublishingOptions :: Maybe (HashMap LogType LogPublishingOption)

Map of LogType and LogPublishingOption, each containing options to publish a given type of OpenSearch log.

clusterConfig :: Maybe ClusterConfig

The type and number of instances to instantiate for the domain cluster.

advancedSecurityOptions :: Maybe AdvancedSecurityOptionsInput

Specifies advanced security options.

snapshotOptions :: Maybe SnapshotOptions

Option to set the time, in UTC format, for the daily automated snapshot. Default value is 0 hours.

cognitoOptions :: Maybe CognitoOptions

Options to specify the Cognito user and identity pools for OpenSearch Dashboards authentication. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

encryptionAtRestOptions :: Maybe EncryptionAtRestOptions

Specifies encryption of data at rest options.

vPCOptions :: Maybe VPCOptions

Options to specify the subnets and security groups for the VPC endpoint. For more information, see Launching your Amazon OpenSearch Service domains using a VPC .

domainEndpointOptions :: Maybe DomainEndpointOptions

Options to specify configuration that will be applied to the domain endpoint.

advancedOptions :: Maybe (HashMap Text Text)

Modifies the advanced option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Advanced options for more information.

domainName :: Text

The name of the domain you're updating.

newUpdateDomainConfig

Arguments

:: Text
$sel:domainName:UpdateDomainConfig'

-> UpdateDomainConfig

Create a value of UpdateDomainConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eBSOptions:UpdateDomainConfig', updateDomainConfig_eBSOptions - Specify the type and size of the EBS volume to use.

$sel:nodeToNodeEncryptionOptions:UpdateDomainConfig', updateDomainConfig_nodeToNodeEncryptionOptions - Specifies node-to-node encryption options.

$sel:accessPolicies:UpdateDomainConfig', updateDomainConfig_accessPolicies - IAM access policy as a JSON-formatted string.

$sel:autoTuneOptions:UpdateDomainConfig', updateDomainConfig_autoTuneOptions - Specifies Auto-Tune options.

$sel:logPublishingOptions:UpdateDomainConfig', updateDomainConfig_logPublishingOptions - Map of LogType and LogPublishingOption, each containing options to publish a given type of OpenSearch log.

$sel:clusterConfig:UpdateDomainConfig', updateDomainConfig_clusterConfig - The type and number of instances to instantiate for the domain cluster.

$sel:advancedSecurityOptions:UpdateDomainConfig', updateDomainConfig_advancedSecurityOptions - Specifies advanced security options.

$sel:snapshotOptions:UpdateDomainConfig', updateDomainConfig_snapshotOptions - Option to set the time, in UTC format, for the daily automated snapshot. Default value is 0 hours.

$sel:cognitoOptions:UpdateDomainConfig', updateDomainConfig_cognitoOptions - Options to specify the Cognito user and identity pools for OpenSearch Dashboards authentication. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

$sel:encryptionAtRestOptions:UpdateDomainConfig', updateDomainConfig_encryptionAtRestOptions - Specifies encryption of data at rest options.

$sel:vPCOptions:UpdateDomainConfig', updateDomainConfig_vPCOptions - Options to specify the subnets and security groups for the VPC endpoint. For more information, see Launching your Amazon OpenSearch Service domains using a VPC .

$sel:domainEndpointOptions:UpdateDomainConfig', updateDomainConfig_domainEndpointOptions - Options to specify configuration that will be applied to the domain endpoint.

$sel:advancedOptions:UpdateDomainConfig', updateDomainConfig_advancedOptions - Modifies the advanced option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Advanced options for more information.

$sel:domainName:UpdateDomainConfig', updateDomainConfig_domainName - The name of the domain you're updating.

Request Lenses

updateDomainConfig_eBSOptions :: Lens' UpdateDomainConfig (Maybe EBSOptions)

Specify the type and size of the EBS volume to use.

updateDomainConfig_nodeToNodeEncryptionOptions :: Lens' UpdateDomainConfig (Maybe NodeToNodeEncryptionOptions)

Specifies node-to-node encryption options.

updateDomainConfig_accessPolicies :: Lens' UpdateDomainConfig (Maybe Text)

IAM access policy as a JSON-formatted string.

updateDomainConfig_autoTuneOptions :: Lens' UpdateDomainConfig (Maybe AutoTuneOptions)

Specifies Auto-Tune options.

updateDomainConfig_logPublishingOptions :: Lens' UpdateDomainConfig (Maybe (HashMap LogType LogPublishingOption))

Map of LogType and LogPublishingOption, each containing options to publish a given type of OpenSearch log.

updateDomainConfig_clusterConfig :: Lens' UpdateDomainConfig (Maybe ClusterConfig)

The type and number of instances to instantiate for the domain cluster.

updateDomainConfig_advancedSecurityOptions :: Lens' UpdateDomainConfig (Maybe AdvancedSecurityOptionsInput)

Specifies advanced security options.

updateDomainConfig_snapshotOptions :: Lens' UpdateDomainConfig (Maybe SnapshotOptions)

Option to set the time, in UTC format, for the daily automated snapshot. Default value is 0 hours.

updateDomainConfig_cognitoOptions :: Lens' UpdateDomainConfig (Maybe CognitoOptions)

Options to specify the Cognito user and identity pools for OpenSearch Dashboards authentication. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

updateDomainConfig_encryptionAtRestOptions :: Lens' UpdateDomainConfig (Maybe EncryptionAtRestOptions)

Specifies encryption of data at rest options.

updateDomainConfig_vPCOptions :: Lens' UpdateDomainConfig (Maybe VPCOptions)

Options to specify the subnets and security groups for the VPC endpoint. For more information, see Launching your Amazon OpenSearch Service domains using a VPC .

updateDomainConfig_domainEndpointOptions :: Lens' UpdateDomainConfig (Maybe DomainEndpointOptions)

Options to specify configuration that will be applied to the domain endpoint.

updateDomainConfig_advancedOptions :: Lens' UpdateDomainConfig (Maybe (HashMap Text Text))

Modifies the advanced option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Advanced options for more information.

updateDomainConfig_domainName :: Lens' UpdateDomainConfig Text

The name of the domain you're updating.

Destructuring the Response

data UpdateDomainConfigResponse

The result of an UpdateDomain request. Contains the status of the domain being updated.

See: newUpdateDomainConfigResponse smart constructor.

Constructors

UpdateDomainConfigResponse'

Fields

httpStatus :: Int

The response's http status code.

domainConfig :: DomainConfig

The status of the updated domain.

newUpdateDomainConfigResponse

Arguments

:: Int
$sel:httpStatus:UpdateDomainConfigResponse'

-> DomainConfig
$sel:domainConfig:UpdateDomainConfigResponse'

-> UpdateDomainConfigResponse

Create a value of UpdateDomainConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDomainConfigResponse', updateDomainConfigResponse_httpStatus - The response's http status code.

$sel:domainConfig:UpdateDomainConfigResponse', updateDomainConfigResponse_domainConfig - The status of the updated domain.

Response Lenses

updateDomainConfigResponse_httpStatus :: Lens' UpdateDomainConfigResponse Int

The response's http status code.

updateDomainConfigResponse_domainConfig :: Lens' UpdateDomainConfigResponse DomainConfig

The status of the updated domain.