Amazonka.OpenSearch.UpdatePackage

Description

Updates a package for use with Amazon OpenSearch Service domains.

Creating a Request

data UpdatePackage

Container for request parameters to the UpdatePackage operation.

See: newUpdatePackage smart constructor.

Constructors

UpdatePackage'

Fields

packageDescription :: Maybe Text

A new description of the package.

commitMessage :: Maybe Text

A commit message for the new version which is shown as part of GetPackageVersionHistoryResponse.

packageID :: Text

The unique identifier for the package.

packageSource :: PackageSource

newUpdatePackage

Arguments

:: Text
$sel:packageID:UpdatePackage'

-> PackageSource
$sel:packageSource:UpdatePackage'

-> UpdatePackage

Create a value of UpdatePackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageDescription:UpdatePackage', updatePackage_packageDescription - A new description of the package.

$sel:commitMessage:UpdatePackage', updatePackage_commitMessage - A commit message for the new version which is shown as part of GetPackageVersionHistoryResponse.

$sel:packageID:UpdatePackage', updatePackage_packageID - The unique identifier for the package.

$sel:packageSource:UpdatePackage', updatePackage_packageSource - Undocumented member.

Request Lenses

updatePackage_packageDescription :: Lens' UpdatePackage (Maybe Text)

A new description of the package.

updatePackage_commitMessage :: Lens' UpdatePackage (Maybe Text)

A commit message for the new version which is shown as part of GetPackageVersionHistoryResponse.

updatePackage_packageID :: Lens' UpdatePackage Text

The unique identifier for the package.

updatePackage_packageSource :: Lens' UpdatePackage PackageSource

Undocumented member.

Destructuring the Response

data UpdatePackageResponse

Container for the response returned by the UpdatePackage operation.

See: newUpdatePackageResponse smart constructor.

Constructors

UpdatePackageResponse'

Fields

packageDetails :: Maybe PackageDetails

Information about the package.

httpStatus :: Int

The response's http status code.

newUpdatePackageResponse

Arguments

:: Int
$sel:httpStatus:UpdatePackageResponse'

-> UpdatePackageResponse

Create a value of UpdatePackageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageDetails:UpdatePackageResponse', updatePackageResponse_packageDetails - Information about the package.

$sel:httpStatus:UpdatePackageResponse', updatePackageResponse_httpStatus - The response's http status code.

Response Lenses

updatePackageResponse_packageDetails :: Lens' UpdatePackageResponse (Maybe PackageDetails)

Information about the package.

updatePackageResponse_httpStatus :: Lens' UpdatePackageResponse Int

The response's http status code.