Amazonka.OpsWorks.CreateDeployment

Description

Runs deployment or stack commands. For more information, see Deploying Apps and Run Stack Commands.

Required Permissions: To use this action, an IAM user must have a Deploy or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data CreateDeployment

See: newCreateDeployment smart constructor.

Constructors

CreateDeployment'

Fields

customJson :: Maybe Text

A string that contains user-defined, custom JSON. You can use this parameter to override some corresponding default stack configuration JSON values. The string should be in the following format:

"{\"key1\": \"value1\", \"key2\": \"value2\",...}"

For more information about custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes and Overriding Attributes With Custom JSON.

appId :: Maybe Text

The app ID. This parameter is required for app deployments, but not for other deployment commands.

instanceIds :: Maybe [Text]

The instance IDs for the deployment targets.

layerIds :: Maybe [Text]

The layer IDs for the deployment targets.

comment :: Maybe Text

A user-defined comment.

stackId :: Text

The stack ID.

command :: DeploymentCommand

A DeploymentCommand object that specifies the deployment command and any associated arguments.

newCreateDeployment

Arguments

:: Text
$sel:stackId:CreateDeployment'

-> DeploymentCommand
$sel:command:CreateDeployment'

-> CreateDeployment

Create a value of CreateDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customJson:CreateDeployment', createDeployment_customJson - A string that contains user-defined, custom JSON. You can use this parameter to override some corresponding default stack configuration JSON values. The string should be in the following format:

"{\"key1\": \"value1\", \"key2\": \"value2\",...}"

For more information about custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes and Overriding Attributes With Custom JSON.

$sel:appId:CreateDeployment', createDeployment_appId - The app ID. This parameter is required for app deployments, but not for other deployment commands.

$sel:instanceIds:CreateDeployment', createDeployment_instanceIds - The instance IDs for the deployment targets.

$sel:layerIds:CreateDeployment', createDeployment_layerIds - The layer IDs for the deployment targets.

$sel:comment:CreateDeployment', createDeployment_comment - A user-defined comment.

$sel:stackId:CreateDeployment', createDeployment_stackId - The stack ID.

$sel:command:CreateDeployment', createDeployment_command - A DeploymentCommand object that specifies the deployment command and any associated arguments.

Request Lenses

createDeployment_customJson :: Lens' CreateDeployment (Maybe Text)

A string that contains user-defined, custom JSON. You can use this parameter to override some corresponding default stack configuration JSON values. The string should be in the following format:

"{\"key1\": \"value1\", \"key2\": \"value2\",...}"

For more information about custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes and Overriding Attributes With Custom JSON.

createDeployment_appId :: Lens' CreateDeployment (Maybe Text)

The app ID. This parameter is required for app deployments, but not for other deployment commands.

createDeployment_instanceIds :: Lens' CreateDeployment (Maybe [Text])

The instance IDs for the deployment targets.

createDeployment_layerIds :: Lens' CreateDeployment (Maybe [Text])

The layer IDs for the deployment targets.

createDeployment_comment :: Lens' CreateDeployment (Maybe Text)

A user-defined comment.

createDeployment_stackId :: Lens' CreateDeployment Text

The stack ID.

createDeployment_command :: Lens' CreateDeployment DeploymentCommand

A DeploymentCommand object that specifies the deployment command and any associated arguments.

Destructuring the Response

data CreateDeploymentResponse

Contains the response to a CreateDeployment request.

See: newCreateDeploymentResponse smart constructor.

Constructors

CreateDeploymentResponse'

Fields

deploymentId :: Maybe Text

The deployment ID, which can be used with other requests to identify the deployment.

httpStatus :: Int

The response's http status code.

newCreateDeploymentResponse

Arguments

:: Int
$sel:httpStatus:CreateDeploymentResponse'

-> CreateDeploymentResponse

Create a value of CreateDeploymentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:CreateDeploymentResponse', createDeploymentResponse_deploymentId - The deployment ID, which can be used with other requests to identify the deployment.

$sel:httpStatus:CreateDeploymentResponse', createDeploymentResponse_httpStatus - The response's http status code.

Response Lenses

createDeploymentResponse_deploymentId :: Lens' CreateDeploymentResponse (Maybe Text)

The deployment ID, which can be used with other requests to identify the deployment.

createDeploymentResponse_httpStatus :: Lens' CreateDeploymentResponse Int

The response's http status code.