Amazonka.OpsWorks.DescribeAgentVersions

Description

Describes the available AWS OpsWorks Stacks agent versions. You must specify a stack ID or a configuration manager. DescribeAgentVersions returns a list of available agent versions for the specified stack or configuration manager.

Creating a Request

data DescribeAgentVersions

See: newDescribeAgentVersions smart constructor.

Constructors

DescribeAgentVersions'

Fields

configurationManager :: Maybe StackConfigurationManager

The configuration manager.

stackId :: Maybe Text

The stack ID.

newDescribeAgentVersions :: DescribeAgentVersions

Create a value of DescribeAgentVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationManager:DescribeAgentVersions', describeAgentVersions_configurationManager - The configuration manager.

$sel:stackId:DescribeAgentVersions', describeAgentVersions_stackId - The stack ID.

Request Lenses

describeAgentVersions_configurationManager :: Lens' DescribeAgentVersions (Maybe StackConfigurationManager)

The configuration manager.

describeAgentVersions_stackId :: Lens' DescribeAgentVersions (Maybe Text)

The stack ID.

Destructuring the Response

data DescribeAgentVersionsResponse

Contains the response to a DescribeAgentVersions request.

See: newDescribeAgentVersionsResponse smart constructor.

Constructors

DescribeAgentVersionsResponse'

Fields

agentVersions :: Maybe [AgentVersion]

The agent versions for the specified stack or configuration manager. Note that this value is the complete version number, not the abbreviated number used by the console.

httpStatus :: Int

The response's http status code.

newDescribeAgentVersionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAgentVersionsResponse'

-> DescribeAgentVersionsResponse

Create a value of DescribeAgentVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentVersions:DescribeAgentVersionsResponse', describeAgentVersionsResponse_agentVersions - The agent versions for the specified stack or configuration manager. Note that this value is the complete version number, not the abbreviated number used by the console.

$sel:httpStatus:DescribeAgentVersionsResponse', describeAgentVersionsResponse_httpStatus - The response's http status code.

Response Lenses

describeAgentVersionsResponse_agentVersions :: Lens' DescribeAgentVersionsResponse (Maybe [AgentVersion])

The agent versions for the specified stack or configuration manager. Note that this value is the complete version number, not the abbreviated number used by the console.

describeAgentVersionsResponse_httpStatus :: Lens' DescribeAgentVersionsResponse Int

The response's http status code.