Amazonka.OpsWorks.DescribeTimeBasedAutoScaling

Description

Describes time-based auto scaling configurations for specified instances.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information about user permissions, see Managing User Permissions.

Creating a Request

data DescribeTimeBasedAutoScaling

See: newDescribeTimeBasedAutoScaling smart constructor.

Constructors

DescribeTimeBasedAutoScaling'

Fields

instanceIds :: [Text]

An array of instance IDs.

newDescribeTimeBasedAutoScaling :: DescribeTimeBasedAutoScaling

Create a value of DescribeTimeBasedAutoScaling with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceIds:DescribeTimeBasedAutoScaling', describeTimeBasedAutoScaling_instanceIds - An array of instance IDs.

Request Lenses

describeTimeBasedAutoScaling_instanceIds :: Lens' DescribeTimeBasedAutoScaling [Text]

An array of instance IDs.

Destructuring the Response

data DescribeTimeBasedAutoScalingResponse

Contains the response to a DescribeTimeBasedAutoScaling request.

See: newDescribeTimeBasedAutoScalingResponse smart constructor.

Constructors

DescribeTimeBasedAutoScalingResponse'

Fields

timeBasedAutoScalingConfigurations :: Maybe [TimeBasedAutoScalingConfiguration]

An array of TimeBasedAutoScalingConfiguration objects that describe the configuration for the specified instances.

httpStatus :: Int

The response's http status code.

newDescribeTimeBasedAutoScalingResponse

Arguments

:: Int
$sel:httpStatus:DescribeTimeBasedAutoScalingResponse'

-> DescribeTimeBasedAutoScalingResponse

Create a value of DescribeTimeBasedAutoScalingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeBasedAutoScalingConfigurations:DescribeTimeBasedAutoScalingResponse', describeTimeBasedAutoScalingResponse_timeBasedAutoScalingConfigurations - An array of TimeBasedAutoScalingConfiguration objects that describe the configuration for the specified instances.

$sel:httpStatus:DescribeTimeBasedAutoScalingResponse', describeTimeBasedAutoScalingResponse_httpStatus - The response's http status code.

Response Lenses

describeTimeBasedAutoScalingResponse_timeBasedAutoScalingConfigurations :: Lens' DescribeTimeBasedAutoScalingResponse (Maybe [TimeBasedAutoScalingConfiguration])

An array of TimeBasedAutoScalingConfiguration objects that describe the configuration for the specified instances.

describeTimeBasedAutoScalingResponse_httpStatus :: Lens' DescribeTimeBasedAutoScalingResponse Int

The response's http status code.