Amazonka.OpsWorks.RegisterInstance

Description

Registers instances that were created outside of AWS OpsWorks Stacks with a specified stack.

We do not recommend using this action to register instances. The complete registration operation includes two tasks: installing the AWS OpsWorks Stacks agent on the instance, and registering the instance with the stack. RegisterInstance handles only the second step. You should instead use the AWS CLI register command, which performs the entire registration operation. For more information, see Registering an Instance with an AWS OpsWorks Stacks Stack.

Registered instances have the same requirements as instances that are created by using the CreateInstance API. For example, registered instances must be running a supported Linux-based operating system, and they must have a supported instance type. For more information about requirements for instances that you want to register, see Preparing the Instance.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data RegisterInstance

See: newRegisterInstance smart constructor.

Constructors

RegisterInstance'

Fields

privateIp :: Maybe Text

The instance's private IP address.

hostname :: Maybe Text

The instance's hostname.

instanceIdentity :: Maybe InstanceIdentity

An InstanceIdentity object that contains the instance's identity.

publicIp :: Maybe Text

The instance's public IP address.

rsaPublicKeyFingerprint :: Maybe Text

The instances public RSA key fingerprint.

rsaPublicKey :: Maybe Text

The instances public RSA key. This key is used to encrypt communication between the instance and the service.

stackId :: Text

The ID of the stack that the instance is to be registered with.

newRegisterInstance

Arguments

:: Text
$sel:stackId:RegisterInstance'

-> RegisterInstance

Create a value of RegisterInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:privateIp:RegisterInstance', registerInstance_privateIp - The instance's private IP address.

$sel:hostname:RegisterInstance', registerInstance_hostname - The instance's hostname.

$sel:instanceIdentity:RegisterInstance', registerInstance_instanceIdentity - An InstanceIdentity object that contains the instance's identity.

$sel:publicIp:RegisterInstance', registerInstance_publicIp - The instance's public IP address.

$sel:rsaPublicKeyFingerprint:RegisterInstance', registerInstance_rsaPublicKeyFingerprint - The instances public RSA key fingerprint.

$sel:rsaPublicKey:RegisterInstance', registerInstance_rsaPublicKey - The instances public RSA key. This key is used to encrypt communication between the instance and the service.

$sel:stackId:RegisterInstance', registerInstance_stackId - The ID of the stack that the instance is to be registered with.

Request Lenses

registerInstance_privateIp :: Lens' RegisterInstance (Maybe Text)

The instance's private IP address.

registerInstance_hostname :: Lens' RegisterInstance (Maybe Text)

The instance's hostname.

registerInstance_instanceIdentity :: Lens' RegisterInstance (Maybe InstanceIdentity)

An InstanceIdentity object that contains the instance's identity.

registerInstance_publicIp :: Lens' RegisterInstance (Maybe Text)

The instance's public IP address.

registerInstance_rsaPublicKeyFingerprint :: Lens' RegisterInstance (Maybe Text)

The instances public RSA key fingerprint.

registerInstance_rsaPublicKey :: Lens' RegisterInstance (Maybe Text)

The instances public RSA key. This key is used to encrypt communication between the instance and the service.

registerInstance_stackId :: Lens' RegisterInstance Text

The ID of the stack that the instance is to be registered with.

Destructuring the Response

data RegisterInstanceResponse

Contains the response to a RegisterInstanceResult request.

See: newRegisterInstanceResponse smart constructor.

Constructors

RegisterInstanceResponse'

Fields

instanceId :: Maybe Text

The registered instance's AWS OpsWorks Stacks ID.

httpStatus :: Int

The response's http status code.

newRegisterInstanceResponse

Arguments

:: Int
$sel:httpStatus:RegisterInstanceResponse'

-> RegisterInstanceResponse

Create a value of RegisterInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:RegisterInstanceResponse', registerInstanceResponse_instanceId - The registered instance's AWS OpsWorks Stacks ID.

$sel:httpStatus:RegisterInstanceResponse', registerInstanceResponse_httpStatus - The response's http status code.

Response Lenses

registerInstanceResponse_instanceId :: Lens' RegisterInstanceResponse (Maybe Text)

The registered instance's AWS OpsWorks Stacks ID.

registerInstanceResponse_httpStatus :: Lens' RegisterInstanceResponse Int

The response's http status code.