Amazonka.OpsWorks.SetTimeBasedAutoScaling

Description

Specify the time-based auto scaling configuration for a specified instance. For more information, see Managing Load with Time-based and Load-based Instances.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data SetTimeBasedAutoScaling

See: newSetTimeBasedAutoScaling smart constructor.

Constructors

SetTimeBasedAutoScaling'

Fields

autoScalingSchedule :: Maybe WeeklyAutoScalingSchedule

An AutoScalingSchedule with the instance schedule.

instanceId :: Text

The instance ID.

newSetTimeBasedAutoScaling

Arguments

:: Text
$sel:instanceId:SetTimeBasedAutoScaling'

-> SetTimeBasedAutoScaling

Create a value of SetTimeBasedAutoScaling with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoScalingSchedule:SetTimeBasedAutoScaling', setTimeBasedAutoScaling_autoScalingSchedule - An AutoScalingSchedule with the instance schedule.

$sel:instanceId:SetTimeBasedAutoScaling', setTimeBasedAutoScaling_instanceId - The instance ID.

Request Lenses

setTimeBasedAutoScaling_autoScalingSchedule :: Lens' SetTimeBasedAutoScaling (Maybe WeeklyAutoScalingSchedule)

An AutoScalingSchedule with the instance schedule.

setTimeBasedAutoScaling_instanceId :: Lens' SetTimeBasedAutoScaling Text

The instance ID.

Destructuring the Response

data SetTimeBasedAutoScalingResponse

See: newSetTimeBasedAutoScalingResponse smart constructor.

Constructors

SetTimeBasedAutoScalingResponse'

newSetTimeBasedAutoScalingResponse :: SetTimeBasedAutoScalingResponse

Create a value of SetTimeBasedAutoScalingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.