Amazonka.OpsWorks.Types.App

Description

Documentation

data App

A description of the app.

See: newApp smart constructor.

Constructors

App'

Fields

sslConfiguration :: Maybe SslConfiguration

An SslConfiguration object with the SSL configuration.

environment :: Maybe [EnvironmentVariable]

An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instances. For more information, see Environment Variables.

There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variable names, values, and protected flag values - cannot exceed 20 KB. This limit should accommodate most if not all use cases, but if you do exceed it, you will cause an exception (API) with an "Environment: is too large (maximum is 20 KB)" message.

enableSsl :: Maybe Bool

Whether to enable SSL for the app.

createdAt :: Maybe Text

When the app was created.

shortname :: Maybe Text

The app's short name.

dataSources :: Maybe [DataSource]

The app's data sources.

appSource :: Maybe Source

A Source object that describes the app repository.

appId :: Maybe Text

The app ID.

attributes :: Maybe (HashMap AppAttributesKeys Text)

The stack attributes.

name :: Maybe Text

The app name.

type' :: Maybe AppType

The app type.

stackId :: Maybe Text

The app stack ID.

domains :: Maybe [Text]

The app vhost settings with multiple domains separated by commas. For example: 'www.example.com, example.com'

description :: Maybe Text

A description of the app.

newApp :: App

Create a value of App with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sslConfiguration:App', app_sslConfiguration - An SslConfiguration object with the SSL configuration.

$sel:environment:App', app_environment - An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instances. For more information, see Environment Variables.

There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variable names, values, and protected flag values - cannot exceed 20 KB. This limit should accommodate most if not all use cases, but if you do exceed it, you will cause an exception (API) with an "Environment: is too large (maximum is 20 KB)" message.

$sel:enableSsl:App', app_enableSsl - Whether to enable SSL for the app.

$sel:createdAt:App', app_createdAt - When the app was created.

$sel:shortname:App', app_shortname - The app's short name.

$sel:dataSources:App', app_dataSources - The app's data sources.

$sel:appSource:App', app_appSource - A Source object that describes the app repository.

$sel:appId:App', app_appId - The app ID.

$sel:attributes:App', app_attributes - The stack attributes.

$sel:name:App', app_name - The app name.

$sel:type':App', app_type - The app type.

$sel:stackId:App', app_stackId - The app stack ID.

$sel:domains:App', app_domains - The app vhost settings with multiple domains separated by commas. For example: 'www.example.com, example.com'

$sel:description:App', app_description - A description of the app.

app_sslConfiguration :: Lens' App (Maybe SslConfiguration)

An SslConfiguration object with the SSL configuration.

app_environment :: Lens' App (Maybe [EnvironmentVariable])

An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instances. For more information, see Environment Variables.

There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variable names, values, and protected flag values - cannot exceed 20 KB. This limit should accommodate most if not all use cases, but if you do exceed it, you will cause an exception (API) with an "Environment: is too large (maximum is 20 KB)" message.

app_enableSsl :: Lens' App (Maybe Bool)

Whether to enable SSL for the app.

app_createdAt :: Lens' App (Maybe Text)

When the app was created.

app_shortname :: Lens' App (Maybe Text)

The app's short name.

app_dataSources :: Lens' App (Maybe [DataSource])

The app's data sources.

app_appSource :: Lens' App (Maybe Source)

A Source object that describes the app repository.

app_appId :: Lens' App (Maybe Text)

The app ID.

app_attributes :: Lens' App (Maybe (HashMap AppAttributesKeys Text))

The stack attributes.

app_name :: Lens' App (Maybe Text)

The app name.

app_type :: Lens' App (Maybe AppType)

The app type.

app_stackId :: Lens' App (Maybe Text)

The app stack ID.

app_domains :: Lens' App (Maybe [Text])

The app vhost settings with multiple domains separated by commas. For example: 'www.example.com, example.com'

app_description :: Lens' App (Maybe Text)

A description of the app.