Amazonka.OpsWorks.Types.EbsBlockDevice

Description

Documentation

data EbsBlockDevice

Describes an Amazon EBS volume. This data type maps directly to the Amazon EC2 EbsBlockDevice data type.

See: newEbsBlockDevice smart constructor.

Constructors

EbsBlockDevice'

Fields

deleteOnTermination :: Maybe Bool

Whether the volume is deleted on instance termination.

volumeSize :: Maybe Int

The volume size, in GiB. For more information, see EbsBlockDevice.

iops :: Maybe Int

The number of I/O operations per second (IOPS) that the volume supports. For more information, see EbsBlockDevice.

volumeType :: Maybe VolumeType

The volume type. gp2 for General Purpose (SSD) volumes, io1 for Provisioned IOPS (SSD) volumes, st1 for Throughput Optimized hard disk drives (HDD), sc1 for Cold HDD,and standard for Magnetic volumes.

If you specify the io1 volume type, you must also specify a value for the Iops attribute. The maximum ratio of provisioned IOPS to requested volume size (in GiB) is 50:1. AWS uses the default volume size (in GiB) specified in the AMI attributes to set IOPS to 50 x (volume size).

snapshotId :: Maybe Text

The snapshot ID.

newEbsBlockDevice :: EbsBlockDevice

Create a value of EbsBlockDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteOnTermination:EbsBlockDevice', ebsBlockDevice_deleteOnTermination - Whether the volume is deleted on instance termination.

$sel:volumeSize:EbsBlockDevice', ebsBlockDevice_volumeSize - The volume size, in GiB. For more information, see EbsBlockDevice.

$sel:iops:EbsBlockDevice', ebsBlockDevice_iops - The number of I/O operations per second (IOPS) that the volume supports. For more information, see EbsBlockDevice.

$sel:volumeType:EbsBlockDevice', ebsBlockDevice_volumeType - The volume type. gp2 for General Purpose (SSD) volumes, io1 for Provisioned IOPS (SSD) volumes, st1 for Throughput Optimized hard disk drives (HDD), sc1 for Cold HDD,and standard for Magnetic volumes.

If you specify the io1 volume type, you must also specify a value for the Iops attribute. The maximum ratio of provisioned IOPS to requested volume size (in GiB) is 50:1. AWS uses the default volume size (in GiB) specified in the AMI attributes to set IOPS to 50 x (volume size).

$sel:snapshotId:EbsBlockDevice', ebsBlockDevice_snapshotId - The snapshot ID.

ebsBlockDevice_deleteOnTermination :: Lens' EbsBlockDevice (Maybe Bool)

Whether the volume is deleted on instance termination.

ebsBlockDevice_volumeSize :: Lens' EbsBlockDevice (Maybe Int)

The volume size, in GiB. For more information, see EbsBlockDevice.

ebsBlockDevice_iops :: Lens' EbsBlockDevice (Maybe Int)

The number of I/O operations per second (IOPS) that the volume supports. For more information, see EbsBlockDevice.

ebsBlockDevice_volumeType :: Lens' EbsBlockDevice (Maybe VolumeType)

The volume type. gp2 for General Purpose (SSD) volumes, io1 for Provisioned IOPS (SSD) volumes, st1 for Throughput Optimized hard disk drives (HDD), sc1 for Cold HDD,and standard for Magnetic volumes.

If you specify the io1 volume type, you must also specify a value for the Iops attribute. The maximum ratio of provisioned IOPS to requested volume size (in GiB) is 50:1. AWS uses the default volume size (in GiB) specified in the AMI attributes to set IOPS to 50 x (volume size).

ebsBlockDevice_snapshotId :: Lens' EbsBlockDevice (Maybe Text)

The snapshot ID.