Amazonka.OpsWorks.Types.Layer

Description

Documentation

data Layer

Describes a layer.

See: newLayer smart constructor.

Constructors

Layer'

Fields

customInstanceProfileArn :: Maybe Text

The ARN of the default IAM profile to be used for the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

customSecurityGroupIds :: Maybe [Text]

An array containing the layer's custom security group IDs.

installUpdatesOnBoot :: Maybe Bool

Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

cloudWatchLogsConfiguration :: Maybe CloudWatchLogsConfiguration

The Amazon CloudWatch Logs configuration settings for the layer.

lifecycleEventConfiguration :: Maybe LifecycleEventConfiguration

A LifeCycleEventConfiguration object that specifies the Shutdown event configuration.

arn :: Maybe Text

The Amazon Resource Number (ARN) of a layer.

createdAt :: Maybe Text

Date when the layer was created.

shortname :: Maybe Text

The layer short name.

defaultRecipes :: Maybe Recipes

AWS OpsWorks Stacks supports five lifecycle events: setup, configuration, deploy, undeploy, and shutdown. For each layer, AWS OpsWorks Stacks runs a set of standard recipes for each event. You can also provide custom recipes for any or all layers and events. AWS OpsWorks Stacks runs custom event recipes after the standard recipes. LayerCustomRecipes specifies the custom recipes for a particular layer to be run in response to each of the five events.

To specify a recipe, use the cookbook's directory name in the repository followed by two colons and the recipe name, which is the recipe's file name without the .rb extension. For example: phpapp2::dbsetup specifies the dbsetup.rb recipe in the repository's phpapp2 folder.

customRecipes :: Maybe Recipes

A LayerCustomRecipes object that specifies the layer's custom recipes.

customJson :: Maybe Text

A JSON formatted string containing the layer's custom stack configuration and deployment attributes.

volumeConfigurations :: Maybe [VolumeConfiguration]

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

enableAutoHealing :: Maybe Bool

Whether auto healing is disabled for the layer.

packages :: Maybe [Text]

An array of Package objects that describe the layer's packages.

attributes :: Maybe (HashMap LayerAttributesKeys (Maybe Text))

The layer attributes.

For the HaproxyStatsPassword, MysqlRootPassword, and GangliaPassword attributes, AWS OpsWorks Stacks returns *****FILTERED***** instead of the actual value

For an ECS Cluster layer, AWS OpsWorks Stacks the EcsClusterArn attribute is set to the cluster's ARN.

name :: Maybe Text

The layer name.

autoAssignPublicIps :: Maybe Bool

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer.

type' :: Maybe LayerType

The layer type.

useEbsOptimizedInstances :: Maybe Bool

Whether the layer uses Amazon EBS-optimized instances.

stackId :: Maybe Text

The layer stack ID.

layerId :: Maybe Text

The layer ID.

defaultSecurityGroupNames :: Maybe [Text]

An array containing the layer's security group names.

autoAssignElasticIps :: Maybe Bool

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.

newLayer :: Layer

Create a value of Layer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customInstanceProfileArn:Layer', layer_customInstanceProfileArn - The ARN of the default IAM profile to be used for the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

$sel:customSecurityGroupIds:Layer', layer_customSecurityGroupIds - An array containing the layer's custom security group IDs.

$sel:installUpdatesOnBoot:Layer', layer_installUpdatesOnBoot - Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

$sel:cloudWatchLogsConfiguration:Layer', layer_cloudWatchLogsConfiguration - The Amazon CloudWatch Logs configuration settings for the layer.

$sel:lifecycleEventConfiguration:Layer', layer_lifecycleEventConfiguration - A LifeCycleEventConfiguration object that specifies the Shutdown event configuration.

$sel:arn:Layer', layer_arn - The Amazon Resource Number (ARN) of a layer.

$sel:createdAt:Layer', layer_createdAt - Date when the layer was created.

$sel:shortname:Layer', layer_shortname - The layer short name.

$sel:defaultRecipes:Layer', layer_defaultRecipes - AWS OpsWorks Stacks supports five lifecycle events: setup, configuration, deploy, undeploy, and shutdown. For each layer, AWS OpsWorks Stacks runs a set of standard recipes for each event. You can also provide custom recipes for any or all layers and events. AWS OpsWorks Stacks runs custom event recipes after the standard recipes. LayerCustomRecipes specifies the custom recipes for a particular layer to be run in response to each of the five events.

To specify a recipe, use the cookbook's directory name in the repository followed by two colons and the recipe name, which is the recipe's file name without the .rb extension. For example: phpapp2::dbsetup specifies the dbsetup.rb recipe in the repository's phpapp2 folder.

$sel:customRecipes:Layer', layer_customRecipes - A LayerCustomRecipes object that specifies the layer's custom recipes.

$sel:customJson:Layer', layer_customJson - A JSON formatted string containing the layer's custom stack configuration and deployment attributes.

$sel:volumeConfigurations:Layer', layer_volumeConfigurations - A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

$sel:enableAutoHealing:Layer', layer_enableAutoHealing - Whether auto healing is disabled for the layer.

$sel:packages:Layer', layer_packages - An array of Package objects that describe the layer's packages.

$sel:attributes:Layer', layer_attributes - The layer attributes.

For the HaproxyStatsPassword, MysqlRootPassword, and GangliaPassword attributes, AWS OpsWorks Stacks returns *****FILTERED***** instead of the actual value

For an ECS Cluster layer, AWS OpsWorks Stacks the EcsClusterArn attribute is set to the cluster's ARN.

$sel:name:Layer', layer_name - The layer name.

$sel:autoAssignPublicIps:Layer', layer_autoAssignPublicIps - For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer.

$sel:type':Layer', layer_type - The layer type.

$sel:useEbsOptimizedInstances:Layer', layer_useEbsOptimizedInstances - Whether the layer uses Amazon EBS-optimized instances.

$sel:stackId:Layer', layer_stackId - The layer stack ID.

$sel:layerId:Layer', layer_layerId - The layer ID.

$sel:defaultSecurityGroupNames:Layer', layer_defaultSecurityGroupNames - An array containing the layer's security group names.

$sel:autoAssignElasticIps:Layer', layer_autoAssignElasticIps - Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.

layer_customInstanceProfileArn :: Lens' Layer (Maybe Text)

The ARN of the default IAM profile to be used for the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

layer_customSecurityGroupIds :: Lens' Layer (Maybe [Text])

An array containing the layer's custom security group IDs.

layer_installUpdatesOnBoot :: Lens' Layer (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

layer_cloudWatchLogsConfiguration :: Lens' Layer (Maybe CloudWatchLogsConfiguration)

The Amazon CloudWatch Logs configuration settings for the layer.

layer_lifecycleEventConfiguration :: Lens' Layer (Maybe LifecycleEventConfiguration)

A LifeCycleEventConfiguration object that specifies the Shutdown event configuration.

layer_arn :: Lens' Layer (Maybe Text)

The Amazon Resource Number (ARN) of a layer.

layer_createdAt :: Lens' Layer (Maybe Text)

Date when the layer was created.

layer_shortname :: Lens' Layer (Maybe Text)

The layer short name.

layer_defaultRecipes :: Lens' Layer (Maybe Recipes)

AWS OpsWorks Stacks supports five lifecycle events: setup, configuration, deploy, undeploy, and shutdown. For each layer, AWS OpsWorks Stacks runs a set of standard recipes for each event. You can also provide custom recipes for any or all layers and events. AWS OpsWorks Stacks runs custom event recipes after the standard recipes. LayerCustomRecipes specifies the custom recipes for a particular layer to be run in response to each of the five events.

To specify a recipe, use the cookbook's directory name in the repository followed by two colons and the recipe name, which is the recipe's file name without the .rb extension. For example: phpapp2::dbsetup specifies the dbsetup.rb recipe in the repository's phpapp2 folder.

layer_customRecipes :: Lens' Layer (Maybe Recipes)

A LayerCustomRecipes object that specifies the layer's custom recipes.

layer_customJson :: Lens' Layer (Maybe Text)

A JSON formatted string containing the layer's custom stack configuration and deployment attributes.

layer_volumeConfigurations :: Lens' Layer (Maybe [VolumeConfiguration])

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

layer_enableAutoHealing :: Lens' Layer (Maybe Bool)

Whether auto healing is disabled for the layer.

layer_packages :: Lens' Layer (Maybe [Text])

An array of Package objects that describe the layer's packages.

layer_attributes :: Lens' Layer (Maybe (HashMap LayerAttributesKeys (Maybe Text)))

The layer attributes.

For the HaproxyStatsPassword, MysqlRootPassword, and GangliaPassword attributes, AWS OpsWorks Stacks returns *****FILTERED***** instead of the actual value

For an ECS Cluster layer, AWS OpsWorks Stacks the EcsClusterArn attribute is set to the cluster's ARN.

layer_name :: Lens' Layer (Maybe Text)

The layer name.

layer_autoAssignPublicIps :: Lens' Layer (Maybe Bool)

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer.

layer_type :: Lens' Layer (Maybe LayerType)

The layer type.

layer_useEbsOptimizedInstances :: Lens' Layer (Maybe Bool)

Whether the layer uses Amazon EBS-optimized instances.

layer_stackId :: Lens' Layer (Maybe Text)

The layer stack ID.

layer_layerId :: Lens' Layer (Maybe Text)

The layer ID.

layer_defaultSecurityGroupNames :: Lens' Layer (Maybe [Text])

An array containing the layer's security group names.

layer_autoAssignElasticIps :: Lens' Layer (Maybe Bool)

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.