Amazonka.OpsWorks.Types.LoadBasedAutoScalingConfiguration

Description

Documentation

data LoadBasedAutoScalingConfiguration

Describes a layer's load-based auto scaling configuration.

See: newLoadBasedAutoScalingConfiguration smart constructor.

Constructors

LoadBasedAutoScalingConfiguration'

Fields

upScaling :: Maybe AutoScalingThresholds

An AutoScalingThresholds object that describes the upscaling configuration, which defines how and when AWS OpsWorks Stacks increases the number of instances.

enable :: Maybe Bool

Whether load-based auto scaling is enabled for the layer.

downScaling :: Maybe AutoScalingThresholds

An AutoScalingThresholds object that describes the downscaling configuration, which defines how and when AWS OpsWorks Stacks reduces the number of instances.

layerId :: Maybe Text

The layer ID.

newLoadBasedAutoScalingConfiguration :: LoadBasedAutoScalingConfiguration

Create a value of LoadBasedAutoScalingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:upScaling:LoadBasedAutoScalingConfiguration', loadBasedAutoScalingConfiguration_upScaling - An AutoScalingThresholds object that describes the upscaling configuration, which defines how and when AWS OpsWorks Stacks increases the number of instances.

$sel:enable:LoadBasedAutoScalingConfiguration', loadBasedAutoScalingConfiguration_enable - Whether load-based auto scaling is enabled for the layer.

$sel:downScaling:LoadBasedAutoScalingConfiguration', loadBasedAutoScalingConfiguration_downScaling - An AutoScalingThresholds object that describes the downscaling configuration, which defines how and when AWS OpsWorks Stacks reduces the number of instances.

$sel:layerId:LoadBasedAutoScalingConfiguration', loadBasedAutoScalingConfiguration_layerId - The layer ID.

loadBasedAutoScalingConfiguration_upScaling :: Lens' LoadBasedAutoScalingConfiguration (Maybe AutoScalingThresholds)

An AutoScalingThresholds object that describes the upscaling configuration, which defines how and when AWS OpsWorks Stacks increases the number of instances.

loadBasedAutoScalingConfiguration_enable :: Lens' LoadBasedAutoScalingConfiguration (Maybe Bool)

Whether load-based auto scaling is enabled for the layer.

loadBasedAutoScalingConfiguration_downScaling :: Lens' LoadBasedAutoScalingConfiguration (Maybe AutoScalingThresholds)

An AutoScalingThresholds object that describes the downscaling configuration, which defines how and when AWS OpsWorks Stacks reduces the number of instances.

loadBasedAutoScalingConfiguration_layerId :: Lens' LoadBasedAutoScalingConfiguration (Maybe Text)

The layer ID.