Amazonka.OpsWorks.Types.Permission

Description

Documentation

data Permission

Describes stack or user permissions.

See: newPermission smart constructor.

Constructors

Permission'

Fields

iamUserArn :: Maybe Text

The Amazon Resource Name (ARN) for an AWS Identity and Access Management (IAM) role. For more information about IAM ARNs, see Using Identifiers.

allowSudo :: Maybe Bool

Whether the user can use sudo.

stackId :: Maybe Text

A stack ID.

level :: Maybe Text

The user's permission level, which must be the following:

deny

show

deploy

manage

iam_only

For more information on the permissions associated with these levels, see Managing User Permissions

allowSsh :: Maybe Bool

Whether the user can use SSH.

newPermission :: Permission

Create a value of Permission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamUserArn:Permission', permission_iamUserArn - The Amazon Resource Name (ARN) for an AWS Identity and Access Management (IAM) role. For more information about IAM ARNs, see Using Identifiers.

$sel:allowSudo:Permission', permission_allowSudo - Whether the user can use sudo.

$sel:stackId:Permission', permission_stackId - A stack ID.

$sel:level:Permission', permission_level - The user's permission level, which must be the following:

deny

show

deploy

manage

iam_only

For more information on the permissions associated with these levels, see Managing User Permissions

$sel:allowSsh:Permission', permission_allowSsh - Whether the user can use SSH.

permission_iamUserArn :: Lens' Permission (Maybe Text)

The Amazon Resource Name (ARN) for an AWS Identity and Access Management (IAM) role. For more information about IAM ARNs, see Using Identifiers.

permission_allowSudo :: Lens' Permission (Maybe Bool)

Whether the user can use sudo.

permission_stackId :: Lens' Permission (Maybe Text)

A stack ID.

permission_level :: Lens' Permission (Maybe Text)

The user's permission level, which must be the following:

deny

show

deploy

manage

iam_only

For more information on the permissions associated with these levels, see Managing User Permissions

permission_allowSsh :: Lens' Permission (Maybe Bool)

Whether the user can use SSH.