Amazonka.OpsWorks.Types.Recipes

Description

Documentation

data Recipes

AWS OpsWorks Stacks supports five lifecycle events: setup, configuration, deploy, undeploy, and shutdown. For each layer, AWS OpsWorks Stacks runs a set of standard recipes for each event. In addition, you can provide custom recipes for any or all layers and events. AWS OpsWorks Stacks runs custom event recipes after the standard recipes. LayerCustomRecipes specifies the custom recipes for a particular layer to be run in response to each of the five events.

To specify a recipe, use the cookbook's directory name in the repository followed by two colons and the recipe name, which is the recipe's file name without the .rb extension. For example: phpapp2::dbsetup specifies the dbsetup.rb recipe in the repository's phpapp2 folder.

See: newRecipes smart constructor.

Constructors

Recipes'

Fields

setup :: Maybe [Text]

An array of custom recipe names to be run following a setup event.

shutdown :: Maybe [Text]

An array of custom recipe names to be run following a shutdown event.

undeploy :: Maybe [Text]

An array of custom recipe names to be run following a undeploy event.

configure :: Maybe [Text]

An array of custom recipe names to be run following a configure event.

deploy :: Maybe [Text]

An array of custom recipe names to be run following a deploy event.

newRecipes :: Recipes

Create a value of Recipes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:setup:Recipes', recipes_setup - An array of custom recipe names to be run following a setup event.

$sel:shutdown:Recipes', recipes_shutdown - An array of custom recipe names to be run following a shutdown event.

$sel:undeploy:Recipes', recipes_undeploy - An array of custom recipe names to be run following a undeploy event.

$sel:configure:Recipes', recipes_configure - An array of custom recipe names to be run following a configure event.

$sel:deploy:Recipes', recipes_deploy - An array of custom recipe names to be run following a deploy event.

recipes_setup :: Lens' Recipes (Maybe [Text])

An array of custom recipe names to be run following a setup event.

recipes_shutdown :: Lens' Recipes (Maybe [Text])

An array of custom recipe names to be run following a shutdown event.

recipes_undeploy :: Lens' Recipes (Maybe [Text])

An array of custom recipe names to be run following a undeploy event.

recipes_configure :: Lens' Recipes (Maybe [Text])

An array of custom recipe names to be run following a configure event.

recipes_deploy :: Lens' Recipes (Maybe [Text])

An array of custom recipe names to be run following a deploy event.