Amazonka.OpsWorks.Types.ShutdownEventConfiguration

Description

Documentation

data ShutdownEventConfiguration

The Shutdown event configuration.

See: newShutdownEventConfiguration smart constructor.

Constructors

ShutdownEventConfiguration'

Fields

executionTimeout :: Maybe Int

The time, in seconds, that AWS OpsWorks Stacks will wait after triggering a Shutdown event before shutting down an instance.

delayUntilElbConnectionsDrained :: Maybe Bool

Whether to enable Elastic Load Balancing connection draining. For more information, see Connection Draining

newShutdownEventConfiguration :: ShutdownEventConfiguration

Create a value of ShutdownEventConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionTimeout:ShutdownEventConfiguration', shutdownEventConfiguration_executionTimeout - The time, in seconds, that AWS OpsWorks Stacks will wait after triggering a Shutdown event before shutting down an instance.

$sel:delayUntilElbConnectionsDrained:ShutdownEventConfiguration', shutdownEventConfiguration_delayUntilElbConnectionsDrained - Whether to enable Elastic Load Balancing connection draining. For more information, see Connection Draining

shutdownEventConfiguration_executionTimeout :: Lens' ShutdownEventConfiguration (Maybe Int)

The time, in seconds, that AWS OpsWorks Stacks will wait after triggering a Shutdown event before shutting down an instance.

shutdownEventConfiguration_delayUntilElbConnectionsDrained :: Lens' ShutdownEventConfiguration (Maybe Bool)

Whether to enable Elastic Load Balancing connection draining. For more information, see Connection Draining