Amazonka.OpsWorks.Types.StackSummary

Description

Documentation

data StackSummary

Summarizes the number of layers, instances, and apps in a stack.

See: newStackSummary smart constructor.

Constructors

StackSummary'

Fields

arn :: Maybe Text

The stack's ARN.

appsCount :: Maybe Int

The number of apps.

name :: Maybe Text

The stack name.

stackId :: Maybe Text

The stack ID.

layersCount :: Maybe Int

The number of layers.

instancesCount :: Maybe InstancesCount

An InstancesCount object with the number of instances in each status.

newStackSummary :: StackSummary

Create a value of StackSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:StackSummary', stackSummary_arn - The stack's ARN.

$sel:appsCount:StackSummary', stackSummary_appsCount - The number of apps.

$sel:name:StackSummary', stackSummary_name - The stack name.

$sel:stackId:StackSummary', stackSummary_stackId - The stack ID.

$sel:layersCount:StackSummary', stackSummary_layersCount - The number of layers.

$sel:instancesCount:StackSummary', stackSummary_instancesCount - An InstancesCount object with the number of instances in each status.

stackSummary_arn :: Lens' StackSummary (Maybe Text)

The stack's ARN.

stackSummary_appsCount :: Lens' StackSummary (Maybe Int)

The number of apps.

stackSummary_name :: Lens' StackSummary (Maybe Text)

The stack name.

stackSummary_stackId :: Lens' StackSummary (Maybe Text)

The stack ID.

stackSummary_layersCount :: Lens' StackSummary (Maybe Int)

The number of layers.

stackSummary_instancesCount :: Lens' StackSummary (Maybe InstancesCount)

An InstancesCount object with the number of instances in each status.