Amazonka.OpsWorksCM.Lens

Description

Operations

AssociateNode

associateNode_serverName :: Lens' AssociateNode Text

The name of the server with which to associate the node.

associateNode_nodeName :: Lens' AssociateNode Text

The name of the node.

associateNode_engineAttributes :: Lens' AssociateNode [EngineAttribute]

Engine attributes used for associating the node.

Attributes accepted in a AssociateNode request for Chef

CHEF_ORGANIZATION: The Chef organization with which the node is associated. By default only one organization named default can exist.

CHEF_NODE_PUBLIC_KEY: A PEM-formatted public key. This key is required for the chef-client agent to access the Chef API.

Attributes accepted in a AssociateNode request for Puppet

PUPPET_NODE_CSR: A PEM-formatted certificate-signing request (CSR) that is created by the node.

associateNodeResponse_nodeAssociationStatusToken :: Lens' AssociateNodeResponse (Maybe Text)

Contains a token which can be passed to the DescribeNodeAssociationStatus API call to get the status of the association request.

associateNodeResponse_httpStatus :: Lens' AssociateNodeResponse Int

The response's http status code.

UpdateServer

updateServer_disableAutomatedBackup :: Lens' UpdateServer (Maybe Bool)

Setting DisableAutomatedBackup to true disables automated or scheduled backups. Automated backups are enabled by default.

updateServer_preferredMaintenanceWindow :: Lens' UpdateServer (Maybe Text)

Undocumented member.

updateServer_preferredBackupWindow :: Lens' UpdateServer (Maybe Text)

Undocumented member.

updateServer_backupRetentionCount :: Lens' UpdateServer (Maybe Int)

Sets the number of automated backups that you want to keep.

updateServer_serverName :: Lens' UpdateServer Text

The name of the server to update.

updateServerResponse_server :: Lens' UpdateServerResponse (Maybe Server)

Contains the response to a UpdateServer request.

updateServerResponse_httpStatus :: Lens' UpdateServerResponse Int

The response's http status code.

DeleteServer

deleteServer_serverName :: Lens' DeleteServer Text

The ID of the server to delete.

deleteServerResponse_httpStatus :: Lens' DeleteServerResponse Int

The response's http status code.

DeleteBackup

deleteBackup_backupId :: Lens' DeleteBackup Text

The ID of the backup to delete. Run the DescribeBackups command to get a list of backup IDs. Backup IDs are in the format ServerName-yyyyMMddHHmmssSSS.

deleteBackupResponse_httpStatus :: Lens' DeleteBackupResponse Int

The response's http status code.

DescribeEvents

describeEvents_nextToken :: Lens' DescribeEvents (Maybe Text)

NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call DescribeEvents again, and assign the token from the previous results as the value of the nextToken parameter. If there are no more results, the response object's nextToken parameter value is null. Setting a nextToken value that was not returned in your previous results causes an InvalidNextTokenException to occur.

describeEvents_maxResults :: Lens' DescribeEvents (Maybe Natural)

To receive a paginated response, use this parameter to specify the maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

describeEvents_serverName :: Lens' DescribeEvents Text

The name of the server for which you want to view events.

describeEventsResponse_serverEvents :: Lens' DescribeEventsResponse (Maybe [ServerEvent])

Contains the response to a DescribeEvents request.

describeEventsResponse_nextToken :: Lens' DescribeEventsResponse (Maybe Text)

NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call DescribeEvents again, and assign the token from the previous results as the value of the nextToken parameter. If there are no more results, the response object's nextToken parameter value is null. Setting a nextToken value that was not returned in your previous results causes an InvalidNextTokenException to occur.

describeEventsResponse_httpStatus :: Lens' DescribeEventsResponse Int

The response's http status code.

DisassociateNode

disassociateNode_engineAttributes :: Lens' DisassociateNode (Maybe [EngineAttribute])

Engine attributes that are used for disassociating the node. No attributes are required for Puppet.

Attributes required in a DisassociateNode request for Chef

CHEF_ORGANIZATION: The Chef organization with which the node was associated. By default only one organization named default can exist.

disassociateNode_serverName :: Lens' DisassociateNode Text

The name of the server from which to disassociate the node.

disassociateNode_nodeName :: Lens' DisassociateNode Text

The name of the client node.

disassociateNodeResponse_nodeAssociationStatusToken :: Lens' DisassociateNodeResponse (Maybe Text)

Contains a token which can be passed to the DescribeNodeAssociationStatus API call to get the status of the disassociation request.

disassociateNodeResponse_httpStatus :: Lens' DisassociateNodeResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_nextToken :: Lens' ListTagsForResource (Maybe Text)

NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call ListTagsForResource again, and assign the token from the previous results as the value of the nextToken parameter. If there are no more results, the response object's nextToken parameter value is null. Setting a nextToken value that was not returned in your previous results causes an InvalidNextTokenException to occur.

listTagsForResource_maxResults :: Lens' ListTagsForResource (Maybe Natural)

To receive a paginated response, use this parameter to specify the maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Number (ARN) of an AWS OpsWorks for Chef Automate or AWS OpsWorks for Puppet Enterprise server for which you want to show applied tags. For example, arn:aws:opsworks-cm:us-west-2:123456789012:server/test-owcm-server/EXAMPLE-66b0-4196-8274-d1a2bEXAMPLE.

listTagsForResourceResponse_nextToken :: Lens' ListTagsForResourceResponse (Maybe Text)

A token that you can use as the value of NextToken in subsequent calls to the API to show more results.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

Tags that have been applied to the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CreateBackup

createBackup_description :: Lens' CreateBackup (Maybe Text)

A user-defined description of the backup.

createBackup_tags :: Lens' CreateBackup (Maybe [Tag])

A map that contains tag keys and tag values to attach to an AWS OpsWorks-CM server backup.

The key cannot be empty.

The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /

The value can be a maximum 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /

Leading and trailing white spaces are trimmed from both the key and value.

A maximum of 50 user-applied tags is allowed for tag-supported AWS OpsWorks-CM resources.

createBackup_serverName :: Lens' CreateBackup Text

The name of the server that you want to back up.

createBackupResponse_backup :: Lens' CreateBackupResponse (Maybe Backup)

Backup created by request.

createBackupResponse_httpStatus :: Lens' CreateBackupResponse Int

The response's http status code.

UpdateServerEngineAttributes

updateServerEngineAttributes_attributeValue :: Lens' UpdateServerEngineAttributes (Maybe Text)

The value to set for the attribute.

updateServerEngineAttributes_serverName :: Lens' UpdateServerEngineAttributes Text

The name of the server to update.

updateServerEngineAttributes_attributeName :: Lens' UpdateServerEngineAttributes Text

The name of the engine attribute to update.

updateServerEngineAttributesResponse_server :: Lens' UpdateServerEngineAttributesResponse (Maybe Server)

Contains the response to an UpdateServerEngineAttributes request.

updateServerEngineAttributesResponse_httpStatus :: Lens' UpdateServerEngineAttributesResponse Int

The response's http status code.

StartMaintenance

startMaintenance_engineAttributes :: Lens' StartMaintenance (Maybe [EngineAttribute])

Engine attributes that are specific to the server on which you want to run maintenance.

Attributes accepted in a StartMaintenance request for Chef

CHEF_MAJOR_UPGRADE: If a Chef Automate server is eligible for upgrade to Chef Automate 2, add this engine attribute to a StartMaintenance request and set the value to true to upgrade the server to Chef Automate 2. For more information, see Upgrade an AWS OpsWorks for Chef Automate Server to Chef Automate 2.

startMaintenance_serverName :: Lens' StartMaintenance Text

The name of the server on which to run maintenance.

startMaintenanceResponse_server :: Lens' StartMaintenanceResponse (Maybe Server)

Contains the response to a StartMaintenance request.

startMaintenanceResponse_httpStatus :: Lens' StartMaintenanceResponse Int

The response's http status code.

DescribeBackups

describeBackups_serverName :: Lens' DescribeBackups (Maybe Text)

Returns backups for the server with the specified ServerName.

describeBackups_backupId :: Lens' DescribeBackups (Maybe Text)

Describes a single backup.

describeBackups_nextToken :: Lens' DescribeBackups (Maybe Text)

This is not currently implemented for DescribeBackups requests.

describeBackups_maxResults :: Lens' DescribeBackups (Maybe Natural)

This is not currently implemented for DescribeBackups requests.

describeBackupsResponse_backups :: Lens' DescribeBackupsResponse (Maybe [Backup])

Contains the response to a DescribeBackups request.

describeBackupsResponse_nextToken :: Lens' DescribeBackupsResponse (Maybe Text)

This is not currently implemented for DescribeBackups requests.

describeBackupsResponse_httpStatus :: Lens' DescribeBackupsResponse Int

The response's http status code.

CreateServer

createServer_engineVersion :: Lens' CreateServer (Maybe Text)

The major release version of the engine that you want to use. For a Chef server, the valid value for EngineVersion is currently 2. For a Puppet server, valid values are 2019 or 2017.

createServer_disableAutomatedBackup :: Lens' CreateServer (Maybe Bool)

Enable or disable scheduled backups. Valid values are true or false. The default value is true.

createServer_securityGroupIds :: Lens' CreateServer (Maybe [Text])

A list of security group IDs to attach to the Amazon EC2 instance. If you add this parameter, the specified security groups must be within the VPC that is specified by SubnetIds.

If you do not specify this parameter, AWS OpsWorks CM creates one new security group that uses TCP ports 22 and 443, open to 0.0.0.0/0 (everyone).

createServer_associatePublicIpAddress :: Lens' CreateServer (Maybe Bool)

Associate a public IP address with a server that you are launching. Valid values are true or false. The default value is true.

createServer_subnetIds :: Lens' CreateServer (Maybe [Text])

The IDs of subnets in which to launch the server EC2 instance.

Amazon EC2-Classic customers: This field is required. All servers must run within a VPC. The VPC must have "Auto Assign Public IP" enabled.

EC2-VPC customers: This field is optional. If you do not specify subnet IDs, your EC2 instances are created in a default subnet that is selected by Amazon EC2. If you specify subnet IDs, the VPC must have "Auto Assign Public IP" enabled.

For more information about supported Amazon EC2 platforms, see Supported Platforms.

createServer_keyPair :: Lens' CreateServer (Maybe Text)

The Amazon EC2 key pair to set for the instance. This parameter is optional; if desired, you may specify this parameter to connect to your instances by using SSH.

createServer_backupId :: Lens' CreateServer (Maybe Text)

If you specify this field, AWS OpsWorks CM creates the server by using the backup represented by BackupId.

createServer_customDomain :: Lens' CreateServer (Maybe Text)

An optional public endpoint of a server, such as https://aws.my-company.com. To access the server, create a CNAME DNS record in your preferred DNS service that points the custom domain to the endpoint that is generated when the server is created (the value of the CreateServer Endpoint attribute). You cannot access the server by using the generated Endpoint value if the server is using a custom domain. If you specify a custom domain, you must also specify values for CustomCertificate and CustomPrivateKey.

createServer_customPrivateKey :: Lens' CreateServer (Maybe Text)

A private key in PEM format for connecting to the server by using HTTPS. The private key must not be encrypted; it cannot be protected by a password or passphrase. If you specify a custom private key, you must also specify values for CustomDomain and CustomCertificate.

createServer_engineModel :: Lens' CreateServer (Maybe Text)

The engine model of the server. Valid values in this release include Monolithic for Puppet and Single for Chef.

createServer_engineAttributes :: Lens' CreateServer (Maybe [EngineAttribute])

Optional engine attributes on a specified server.

Attributes accepted in a Chef createServer request:

CHEF_AUTOMATE_PIVOTAL_KEY: A base64-encoded RSA public key. The corresponding private key is required to access the Chef API. When no CHEF_AUTOMATE_PIVOTAL_KEY is set, a private key is generated and returned in the response.

CHEF_AUTOMATE_ADMIN_PASSWORD: The password for the administrative user in the Chef Automate web-based dashboard. The password length is a minimum of eight characters, and a maximum of 32. The password can contain letters, numbers, and special characters (!/@#$%^&+=_). The password must contain at least one lower case letter, one upper case letter, one number, and one special character. When no CHEF_AUTOMATE_ADMIN_PASSWORD is set, one is generated and returned in the response.

Attributes accepted in a Puppet createServer request:

PUPPET_ADMIN_PASSWORD: To work with the Puppet Enterprise console, a password must use ASCII characters.

PUPPET_R10K_REMOTE: The r10k remote is the URL of your control repository (for example, ssh://git@your.git-repo.com:user/control-repo.git). Specifying an r10k remote opens TCP port 8170.

PUPPET_R10K_PRIVATE_KEY: If you are using a private Git repository, add PUPPET_R10K_PRIVATE_KEY to specify a PEM-encoded private SSH key.

createServer_preferredMaintenanceWindow :: Lens' CreateServer (Maybe Text)

The start time for a one-hour period each week during which AWS OpsWorks CM performs maintenance on the instance. Valid values must be specified in the following format: DDD:HH:MM. MM must be specified as 00. The specified time is in coordinated universal time (UTC). The default value is a random one-hour period on Tuesday, Wednesday, or Friday. See TimeWindowDefinition for more information.

Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

createServer_preferredBackupWindow :: Lens' CreateServer (Maybe Text)

The start time for a one-hour period during which AWS OpsWorks CM backs up application-level data on your server if automated backups are enabled. Valid values must be specified in one of the following formats:

HH:MM for daily backups

DDD:HH:MM for weekly backups

MM must be specified as 00. The specified time is in coordinated universal time (UTC). The default value is a random, daily start time.

Example: 08:00, which represents a daily start time of 08:00 UTC.

Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

createServer_customCertificate :: Lens' CreateServer (Maybe Text)

A PEM-formatted HTTPS certificate. The value can be be a single, self-signed certificate, or a certificate chain. If you specify a custom certificate, you must also specify values for CustomDomain and CustomPrivateKey. The following are requirements for the CustomCertificate value:

You can provide either a self-signed, custom certificate, or the full certificate chain.

The certificate must be a valid X509 certificate, or a certificate chain in PEM format.

The certificate must be valid at the time of upload. A certificate can't be used before its validity period begins (the certificate's NotBefore date), or after it expires (the certificate's NotAfter date).

The certificate’s common name or subject alternative names (SANs), if present, must match the value of CustomDomain.

The certificate must match the value of CustomPrivateKey.

createServer_tags :: Lens' CreateServer (Maybe [Tag])

A map that contains tag keys and tag values to attach to an AWS OpsWorks for Chef Automate or AWS OpsWorks for Puppet Enterprise server.

The key cannot be empty.

The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / @

The value can be a maximum 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / @

Leading and trailing white spaces are trimmed from both the key and value.

A maximum of 50 user-applied tags is allowed for any AWS OpsWorks-CM server.

createServer_backupRetentionCount :: Lens' CreateServer (Maybe Natural)

The number of automated backups that you want to keep. Whenever a new backup is created, AWS OpsWorks CM deletes the oldest backups if this number is exceeded. The default value is 1.

createServer_engine :: Lens' CreateServer Text

The configuration management engine to use. Valid values include ChefAutomate and Puppet.

createServer_serverName :: Lens' CreateServer Text

The name of the server. The server name must be unique within your AWS account, within each region. Server names must start with a letter; then letters, numbers, or hyphens (-) are allowed, up to a maximum of 40 characters.

createServer_instanceProfileArn :: Lens' CreateServer Text

The ARN of the instance profile that your Amazon EC2 instances use. Although the AWS OpsWorks console typically creates the instance profile for you, if you are using API commands instead, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.yaml. This template creates a CloudFormation stack that includes the instance profile you need.

createServer_instanceType :: Lens' CreateServer Text

The Amazon EC2 instance type to use. For example, m5.large.

createServer_serviceRoleArn :: Lens' CreateServer Text

The service role that the AWS OpsWorks CM service backend uses to work with your account. Although the AWS OpsWorks management console typically creates the service role for you, if you are using the AWS CLI or API commands, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.yaml. This template creates a CloudFormation stack that includes the service role and instance profile that you need.

createServerResponse_server :: Lens' CreateServerResponse (Maybe Server)

The server that is created by the request.

createServerResponse_httpStatus :: Lens' CreateServerResponse Int

The response's http status code.

RestoreServer

restoreServer_keyPair :: Lens' RestoreServer (Maybe Text)

The name of the key pair to set on the new EC2 instance. This can be helpful if the administrator no longer has the SSH key.

restoreServer_instanceType :: Lens' RestoreServer (Maybe Text)

The type of instance to restore. Valid values must be specified in the following format: ^([cm][34]|t2).* For example, m5.large. Valid values are m5.large, r5.xlarge, and r5.2xlarge. If you do not specify this parameter, RestoreServer uses the instance type from the specified backup.

restoreServer_backupId :: Lens' RestoreServer Text

The ID of the backup that you want to use to restore a server.

restoreServer_serverName :: Lens' RestoreServer Text

The name of the server that you want to restore.

restoreServerResponse_server :: Lens' RestoreServerResponse (Maybe Server)

Undocumented member.

restoreServerResponse_httpStatus :: Lens' RestoreServerResponse Int

The response's http status code.

DescribeNodeAssociationStatus

describeNodeAssociationStatus_nodeAssociationStatusToken :: Lens' DescribeNodeAssociationStatus Text

The token returned in either the AssociateNodeResponse or the DisassociateNodeResponse.

describeNodeAssociationStatus_serverName :: Lens' DescribeNodeAssociationStatus Text

The name of the server from which to disassociate the node.

describeNodeAssociationStatusResponse_engineAttributes :: Lens' DescribeNodeAssociationStatusResponse (Maybe [EngineAttribute])

Attributes specific to the node association. In Puppet, the attibute PUPPET_NODE_CERT contains the signed certificate (the result of the CSR).

describeNodeAssociationStatusResponse_httpStatus :: Lens' DescribeNodeAssociationStatusResponse Int

The response's http status code.

describeNodeAssociationStatusResponse_nodeAssociationStatus :: Lens' DescribeNodeAssociationStatusResponse NodeAssociationStatus

The status of the association or disassociation request.

Possible values:

SUCCESS: The association or disassociation succeeded.

FAILED: The association or disassociation failed.

IN_PROGRESS: The association or disassociation is still in progress.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Number (ARN) of a resource to which you want to apply tags. For example, arn:aws:opsworks-cm:us-west-2:123456789012:server/test-owcm-server/EXAMPLE-66b0-4196-8274-d1a2bEXAMPLE.

tagResource_tags :: Lens' TagResource [Tag]

A map that contains tag keys and tag values to attach to AWS OpsWorks-CM servers or backups.

The key cannot be empty.

The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /

The value can be a maximum 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /

Leading and trailing white spaces are trimmed from both the key and value.

A maximum of 50 user-applied tags is allowed for any AWS OpsWorks-CM server or backup.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

DescribeAccountAttributes

describeAccountAttributesResponse_attributes :: Lens' DescribeAccountAttributesResponse (Maybe [AccountAttribute])

The attributes that are currently set for the account.

describeAccountAttributesResponse_httpStatus :: Lens' DescribeAccountAttributesResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Number (ARN) of a resource from which you want to remove tags. For example, arn:aws:opsworks-cm:us-west-2:123456789012:server/test-owcm-server/EXAMPLE-66b0-4196-8274-d1a2bEXAMPLE.

untagResource_tagKeys :: Lens' UntagResource [Text]

The keys of tags that you want to remove.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DescribeServers

describeServers_serverName :: Lens' DescribeServers (Maybe Text)

Describes the server with the specified ServerName.

describeServers_nextToken :: Lens' DescribeServers (Maybe Text)

This is not currently implemented for DescribeServers requests.

describeServers_maxResults :: Lens' DescribeServers (Maybe Natural)

This is not currently implemented for DescribeServers requests.

describeServersResponse_servers :: Lens' DescribeServersResponse (Maybe [Server])

Contains the response to a DescribeServers request.

For Chef Automate servers: If DescribeServersResponse$Servers$EngineAttributes includes CHEF_MAJOR_UPGRADE_AVAILABLE, you can upgrade the Chef Automate server to Chef Automate 2. To be eligible for upgrade, a server running Chef Automate 1 must have had at least one successful maintenance run after November 1, 2019.

For Puppet servers: DescribeServersResponse$Servers$EngineAttributes contains the following two responses:

PUPPET_API_CA_CERT, the PEM-encoded CA certificate that is used by the Puppet API over TCP port number 8140. The CA certificate is also used to sign node certificates.

PUPPET_API_CRL, a certificate revocation list. The certificate revocation list is for internal maintenance purposes only. For more information about the Puppet certificate revocation list, see Man Page: puppet certificate_revocation_list in the Puppet documentation.

describeServersResponse_nextToken :: Lens' DescribeServersResponse (Maybe Text)

This is not currently implemented for DescribeServers requests.

describeServersResponse_httpStatus :: Lens' DescribeServersResponse Int

The response's http status code.

ExportServerEngineAttribute

exportServerEngineAttribute_inputAttributes :: Lens' ExportServerEngineAttribute (Maybe [EngineAttribute])

The list of engine attributes. The list type is EngineAttribute. An EngineAttribute list item is a pair that includes an attribute name and its value. For the Userdata ExportAttributeName, the following are supported engine attribute names.

RunList In Chef, a list of roles or recipes that are run in the specified order. In Puppet, this parameter is ignored.

OrganizationName In Chef, an organization name. AWS OpsWorks for Chef Automate always creates the organization default. In Puppet, this parameter is ignored.

NodeEnvironment In Chef, a node environment (for example, development, staging, or one-box). In Puppet, this parameter is ignored.

NodeClientVersion In Chef, the version of the Chef engine (three numbers separated by dots, such as 13.8.5). If this attribute is empty, OpsWorks for Chef Automate uses the most current version. In Puppet, this parameter is ignored.

exportServerEngineAttribute_exportAttributeName :: Lens' ExportServerEngineAttribute Text

The name of the export attribute. Currently, the supported export attribute is Userdata. This exports a user data script that includes parameters and values provided in the InputAttributes list.

exportServerEngineAttribute_serverName :: Lens' ExportServerEngineAttribute Text

The name of the server from which you are exporting the attribute.

exportServerEngineAttributeResponse_serverName :: Lens' ExportServerEngineAttributeResponse (Maybe Text)

The server name used in the request.

exportServerEngineAttributeResponse_engineAttribute :: Lens' ExportServerEngineAttributeResponse (Maybe EngineAttribute)

The requested engine attribute pair with attribute name and value.

exportServerEngineAttributeResponse_httpStatus :: Lens' ExportServerEngineAttributeResponse Int

The response's http status code.

Types

AccountAttribute

accountAttribute_used :: Lens' AccountAttribute (Maybe Int)

The current usage, such as the current number of servers that are associated with the account.

accountAttribute_maximum :: Lens' AccountAttribute (Maybe Int)

The maximum allowed value.

accountAttribute_name :: Lens' AccountAttribute (Maybe Text)

The attribute name. The following are supported attribute names.

ServerLimit: The number of current servers/maximum number of servers allowed. By default, you can have a maximum of 10 servers.

ManualBackupLimit: The number of current manual backups/maximum number of backups allowed. By default, you can have a maximum of 50 manual backups saved.

Backup

backup_engineVersion :: Lens' Backup (Maybe Text)

The engine version that is obtained from the server when the backup is created.

backup_serviceRoleArn :: Lens' Backup (Maybe Text)

The service role ARN that is obtained from the server when the backup is created.

backup_status :: Lens' Backup (Maybe BackupStatus)

The status of a backup while in progress.

backup_instanceProfileArn :: Lens' Backup (Maybe Text)

The EC2 instance profile ARN that is obtained from the server when the backup is created. Because this value is stored, you are not required to provide the InstanceProfileArn again if you restore a backup.

backup_securityGroupIds :: Lens' Backup (Maybe [Text])

The security group IDs that are obtained from the server when the backup is created.

backup_statusDescription :: Lens' Backup (Maybe Text)

An informational message about backup status.

backup_serverName :: Lens' Backup (Maybe Text)

The name of the server from which the backup was made.

backup_subnetIds :: Lens' Backup (Maybe [Text])

The subnet IDs that are obtained from the server when the backup is created.

backup_keyPair :: Lens' Backup (Maybe Text)

The key pair that is obtained from the server when the backup is created.

backup_createdAt :: Lens' Backup (Maybe UTCTime)

The time stamp when the backup was created in the database. Example: 2016-07-29T13:38:47.520Z

backup_backupId :: Lens' Backup (Maybe Text)

The generated ID of the backup. Example: myServerName-yyyyMMddHHmmssSSS

backup_engine :: Lens' Backup (Maybe Text)

The engine type that is obtained from the server when the backup is created.

backup_instanceType :: Lens' Backup (Maybe Text)

The instance type that is obtained from the server when the backup is created.

backup_engineModel :: Lens' Backup (Maybe Text)

The engine model that is obtained from the server when the backup is created.

backup_preferredMaintenanceWindow :: Lens' Backup (Maybe Text)

The preferred maintenance period that is obtained from the server when the backup is created.

backup_userArn :: Lens' Backup (Maybe Text)

The IAM user ARN of the requester for manual backups. This field is empty for automated backups.

backup_preferredBackupWindow :: Lens' Backup (Maybe Text)

The preferred backup period that is obtained from the server when the backup is created.

backup_s3LogUrl :: Lens' Backup (Maybe Text)

The Amazon S3 URL of the backup's log file.

backup_s3DataSize :: Lens' Backup (Maybe Int)

This field is deprecated and is no longer used.

backup_backupArn :: Lens' Backup (Maybe Text)

The ARN of the backup.

backup_s3DataUrl :: Lens' Backup (Maybe Text)

This field is deprecated and is no longer used.

backup_description :: Lens' Backup (Maybe Text)

A user-provided description for a manual backup. This field is empty for automated backups.

backup_backupType :: Lens' Backup (Maybe BackupType)

The backup type. Valid values are automated or manual.

backup_toolsVersion :: Lens' Backup (Maybe Text)

The version of AWS OpsWorks CM-specific tools that is obtained from the server when the backup is created.

EngineAttribute

engineAttribute_value :: Lens' EngineAttribute (Maybe Text)

The value of the engine attribute.

engineAttribute_name :: Lens' EngineAttribute (Maybe Text)

The name of the engine attribute.

Server

server_engineVersion :: Lens' Server (Maybe Text)

The engine version of the server. For a Chef server, the valid value for EngineVersion is currently 2. For a Puppet server, specify either 2019 or 2017.

server_serviceRoleArn :: Lens' Server (Maybe Text)

The service role ARN used to create the server.

server_disableAutomatedBackup :: Lens' Server (Maybe Bool)

Disables automated backups. The number of stored backups is dependent on the value of PreferredBackupCount.

server_status :: Lens' Server (Maybe ServerStatus)

The server's status. This field displays the states of actions in progress, such as creating, running, or backing up the server, as well as the server's health state.

server_instanceProfileArn :: Lens' Server (Maybe Text)

The instance profile ARN of the server.

server_securityGroupIds :: Lens' Server (Maybe [Text])

The security group IDs for the server, as specified in the CloudFormation stack. These might not be the same security groups that are shown in the EC2 console.

server_associatePublicIpAddress :: Lens' Server (Maybe Bool)

Associate a public IP address with a server that you are launching.

server_serverName :: Lens' Server (Maybe Text)

The name of the server.

server_subnetIds :: Lens' Server (Maybe [Text])

The subnet IDs specified in a CreateServer request.

server_keyPair :: Lens' Server (Maybe Text)

The key pair associated with the server.

server_createdAt :: Lens' Server (Maybe UTCTime)

Time stamp of server creation. Example 2016-07-29T13:38:47.520Z

server_serverArn :: Lens' Server (Maybe Text)

The ARN of the server.

server_customDomain :: Lens' Server (Maybe Text)

An optional public endpoint of a server, such as https://aws.my-company.com. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

server_engine :: Lens' Server (Maybe Text)

The engine type of the server. Valid values in this release include ChefAutomate and Puppet.

server_maintenanceStatus :: Lens' Server (Maybe MaintenanceStatus)

The status of the most recent server maintenance run. Shows SUCCESS or FAILED.

server_instanceType :: Lens' Server (Maybe Text)

The instance type for the server, as specified in the CloudFormation stack. This might not be the same instance type that is shown in the EC2 console.

server_engineModel :: Lens' Server (Maybe Text)

The engine model of the server. Valid values in this release include Monolithic for Puppet and Single for Chef.

server_engineAttributes :: Lens' Server (Maybe [EngineAttribute])

The response of a createServer() request returns the master credential to access the server in EngineAttributes. These credentials are not stored by AWS OpsWorks CM; they are returned only as part of the result of createServer().

Attributes returned in a createServer response for Chef

CHEF_AUTOMATE_PIVOTAL_KEY: A base64-encoded RSA private key that is generated by AWS OpsWorks for Chef Automate. This private key is required to access the Chef API.

CHEF_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Chef starter kit, which includes a README, a configuration file, and the required RSA private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents. From this directory, you can run Knife commands.

Attributes returned in a createServer response for Puppet

PUPPET_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Puppet starter kit, including a README and a required private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents.

PUPPET_ADMIN_PASSWORD: An administrator password that you can use to sign in to the Puppet Enterprise console after the server is online.

server_preferredMaintenanceWindow :: Lens' Server (Maybe Text)

The preferred maintenance period specified for the server.

server_preferredBackupWindow :: Lens' Server (Maybe Text)

The preferred backup period specified for the server.

server_statusReason :: Lens' Server (Maybe Text)

Depending on the server status, this field has either a human-readable message (such as a create or backup error), or an escaped block of JSON (used for health check results).

server_endpoint :: Lens' Server (Maybe Text)

A DNS name that can be used to access the engine. Example: myserver-asdfghjkl.us-east-1.opsworks.io. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

server_cloudFormationStackArn :: Lens' Server (Maybe Text)

The ARN of the CloudFormation stack that was used to create the server.

server_backupRetentionCount :: Lens' Server (Maybe Int)

The number of automated backups to keep.

ServerEvent

serverEvent_logUrl :: Lens' ServerEvent (Maybe Text)

The Amazon S3 URL of the event's log file.

serverEvent_serverName :: Lens' ServerEvent (Maybe Text)

The name of the server on or for which the event occurred.

serverEvent_createdAt :: Lens' ServerEvent (Maybe UTCTime)

The time when the event occurred.

serverEvent_message :: Lens' ServerEvent (Maybe Text)

A human-readable informational or status message.

Tag

tag_key :: Lens' Tag Text

A tag key, such as Stage or Name. A tag key cannot be empty. The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /

tag_value :: Lens' Tag Text

An optional tag value, such as Production or test-owcm-server. The value can be a maximum of 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /