Amazonka.OpsWorksCM.UpdateServer

Description

Updates settings for a server.

This operation is synchronous.

Creating a Request

data UpdateServer

See: newUpdateServer smart constructor.

Constructors

UpdateServer'

Fields

disableAutomatedBackup :: Maybe Bool

Setting DisableAutomatedBackup to true disables automated or scheduled backups. Automated backups are enabled by default.

preferredMaintenanceWindow :: Maybe Text

preferredBackupWindow :: Maybe Text

backupRetentionCount :: Maybe Int

Sets the number of automated backups that you want to keep.

serverName :: Text

The name of the server to update.

newUpdateServer

Arguments

:: Text
$sel:serverName:UpdateServer'

-> UpdateServer

Create a value of UpdateServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disableAutomatedBackup:UpdateServer', updateServer_disableAutomatedBackup - Setting DisableAutomatedBackup to true disables automated or scheduled backups. Automated backups are enabled by default.

$sel:preferredMaintenanceWindow:UpdateServer', updateServer_preferredMaintenanceWindow - Undocumented member.

$sel:preferredBackupWindow:UpdateServer', updateServer_preferredBackupWindow - Undocumented member.

$sel:backupRetentionCount:UpdateServer', updateServer_backupRetentionCount - Sets the number of automated backups that you want to keep.

$sel:serverName:UpdateServer', updateServer_serverName - The name of the server to update.

Request Lenses

updateServer_disableAutomatedBackup :: Lens' UpdateServer (Maybe Bool)

Setting DisableAutomatedBackup to true disables automated or scheduled backups. Automated backups are enabled by default.

updateServer_preferredMaintenanceWindow :: Lens' UpdateServer (Maybe Text)

Undocumented member.

updateServer_preferredBackupWindow :: Lens' UpdateServer (Maybe Text)

Undocumented member.

updateServer_backupRetentionCount :: Lens' UpdateServer (Maybe Int)

Sets the number of automated backups that you want to keep.

updateServer_serverName :: Lens' UpdateServer Text

The name of the server to update.

Destructuring the Response

data UpdateServerResponse

See: newUpdateServerResponse smart constructor.

Constructors

UpdateServerResponse'

Fields

server :: Maybe Server

Contains the response to a UpdateServer request.

httpStatus :: Int

The response's http status code.

newUpdateServerResponse

Arguments

:: Int
$sel:httpStatus:UpdateServerResponse'

-> UpdateServerResponse

Create a value of UpdateServerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:server:UpdateServerResponse', updateServerResponse_server - Contains the response to a UpdateServer request.

$sel:httpStatus:UpdateServerResponse', updateServerResponse_httpStatus - The response's http status code.

Response Lenses

updateServerResponse_server :: Lens' UpdateServerResponse (Maybe Server)

Contains the response to a UpdateServer request.

updateServerResponse_httpStatus :: Lens' UpdateServerResponse Int

The response's http status code.