Amazonka.Organizations.AcceptHandshake

Description

Sends a response to the originator of a handshake agreeing to the action proposed by the handshake request.

This operation can be called only by the following principals when they also have the relevant IAM permissions:

Invitation to join or Approve all features request handshakes: only a principal from the member account.

The user who calls the API for an invitation to join must have the organizations:AcceptHandshake permission. If you enabled all features in the organization, the user must also have the iam:CreateServiceLinkedRole permission so that AWS Organizations can create the required service-linked role named AWSServiceRoleForOrganizations. For more information, see AWS Organizations and Service-Linked Roles in the AWS Organizations User Guide.

Enable all features final confirmation handshake: only a principal from the management account.

For more information about invitations, see Inviting an AWS Account to Join Your Organization in the AWS Organizations User Guide. For more information about requests to enable all features in the organization, see Enabling All Features in Your Organization in the AWS Organizations User Guide.

After you accept a handshake, it continues to appear in the results of relevant APIs for only 30 days. After that, it's deleted.

Creating a Request

data AcceptHandshake

See: newAcceptHandshake smart constructor.

Constructors

AcceptHandshake'

Fields

handshakeId :: Text

The unique identifier (ID) of the handshake that you want to accept.

The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lowercase letters or digits.

newAcceptHandshake

Arguments

:: Text
$sel:handshakeId:AcceptHandshake'

-> AcceptHandshake

Create a value of AcceptHandshake with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:handshakeId:AcceptHandshake', acceptHandshake_handshakeId - The unique identifier (ID) of the handshake that you want to accept.

The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lowercase letters or digits.

Request Lenses

acceptHandshake_handshakeId :: Lens' AcceptHandshake Text

The unique identifier (ID) of the handshake that you want to accept.

The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lowercase letters or digits.

Destructuring the Response

data AcceptHandshakeResponse

See: newAcceptHandshakeResponse smart constructor.

Constructors

AcceptHandshakeResponse'

Fields

handshake :: Maybe Handshake

A structure that contains details about the accepted handshake.

httpStatus :: Int

The response's http status code.

newAcceptHandshakeResponse

Arguments

:: Int
$sel:httpStatus:AcceptHandshakeResponse'

-> AcceptHandshakeResponse

Create a value of AcceptHandshakeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:handshake:AcceptHandshakeResponse', acceptHandshakeResponse_handshake - A structure that contains details about the accepted handshake.

$sel:httpStatus:AcceptHandshakeResponse', acceptHandshakeResponse_httpStatus - The response's http status code.

Response Lenses

acceptHandshakeResponse_handshake :: Lens' AcceptHandshakeResponse (Maybe Handshake)

A structure that contains details about the accepted handshake.

acceptHandshakeResponse_httpStatus :: Lens' AcceptHandshakeResponse Int

The response's http status code.