Amazonka.Organizations.DisableAWSServiceAccess

Description

Disables the integration of an AWS service (the service that is specified by ServicePrincipal) with AWS Organizations. When you disable integration, the specified service no longer can create a service-linked role in new accounts in your organization. This means the service can't perform operations on your behalf on any new accounts in your organization. The service can still perform operations in older accounts until the service completes its clean-up from AWS Organizations.

We strongly recommend that you don't use this command to disable integration between AWS Organizations and the specified AWS service. Instead, use the console or commands that are provided by the specified service. This lets the trusted service perform any required initialization when enabling trusted access, such as creating any required resources and any required clean up of resources when disabling trusted access.

For information about how to disable trusted service access to your organization using the trusted service, see the Learn more link under the Supports Trusted Access column at AWS services that you can use with AWS Organizations. on this page.

If you disable access by using this command, it causes the following actions to occur:

The service can no longer create a service-linked role in the accounts in your organization. This means that the service can't perform operations on your behalf on any new accounts in your organization. The service can still perform operations in older accounts until the service completes its clean-up from AWS Organizations.

The service can no longer perform tasks in the member accounts in the organization, unless those operations are explicitly permitted by the IAM policies that are attached to your roles. This includes any data aggregation from the member accounts to the management account, or to a delegated administrator account, where relevant.

Some services detect this and clean up any remaining data or resources related to the integration, while other services stop accessing the organization but leave any historical data and configuration in place to support a possible re-enabling of the integration.

Using the other service's console or commands to disable the integration ensures that the other service is aware that it can clean up any resources that are required only for the integration. How the service cleans up its resources in the organization's accounts depends on that service. For more information, see the documentation for the other AWS service.

After you perform the DisableAWSServiceAccess operation, the specified service can no longer perform operations in your organization's accounts

For more information about integrating other services with AWS Organizations, including the list of services that work with Organizations, see Integrating AWS Organizations with Other AWS Services in the AWS Organizations User Guide.

This operation can be called only from the organization's management account.

Creating a Request

data DisableAWSServiceAccess

See: newDisableAWSServiceAccess smart constructor.

Constructors

DisableAWSServiceAccess'

Fields

servicePrincipal :: Text

The service principal name of the AWS service for which you want to disable integration with your organization. This is typically in the form of a URL, such as service-abbreviation.amazonaws.com.

newDisableAWSServiceAccess

Arguments

:: Text
$sel:servicePrincipal:DisableAWSServiceAccess'

-> DisableAWSServiceAccess

Create a value of DisableAWSServiceAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:servicePrincipal:DisableAWSServiceAccess', disableAWSServiceAccess_servicePrincipal - The service principal name of the AWS service for which you want to disable integration with your organization. This is typically in the form of a URL, such as service-abbreviation.amazonaws.com.

Request Lenses

disableAWSServiceAccess_servicePrincipal :: Lens' DisableAWSServiceAccess Text

The service principal name of the AWS service for which you want to disable integration with your organization. This is typically in the form of a URL, such as service-abbreviation.amazonaws.com.

Destructuring the Response

data DisableAWSServiceAccessResponse

See: newDisableAWSServiceAccessResponse smart constructor.

Constructors

DisableAWSServiceAccessResponse'

newDisableAWSServiceAccessResponse :: DisableAWSServiceAccessResponse

Create a value of DisableAWSServiceAccessResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.