Amazonka.Organizations.ListOrganizationalUnitsForParent

Description

Lists the organizational units (OUs) in a parent organizational unit or root.

Always check the NextToken response parameter for a null value when calling a List* operation. These operations can occasionally return an empty set of results even when there are more results available. The NextToken response parameter value is null only when there are no more results to display.

This operation can be called only from the organization's management account or by a member account that is a delegated administrator for an AWS service.

This operation returns paginated results.

Creating a Request

data ListOrganizationalUnitsForParent

See: newListOrganizationalUnitsForParent smart constructor.

Constructors

ListOrganizationalUnitsForParent'

Fields

nextToken :: Maybe Text

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

maxResults :: Maybe Natural

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

parentId :: Text

The unique identifier (ID) of the root or OU whose child OUs you want to list.

The regex pattern for a parent ID string requires one of the following:

Root - A string that begins with "r-" followed by from 4 to 32 lowercase letters or digits.

Organizational unit (OU) - A string that begins with "ou-" followed by from 4 to 32 lowercase letters or digits (the ID of the root that the OU is in). This string is followed by a second "-" dash and from 8 to 32 additional lowercase letters or digits.

newListOrganizationalUnitsForParent

Arguments

:: Text
$sel:parentId:ListOrganizationalUnitsForParent'

-> ListOrganizationalUnitsForParent

Create a value of ListOrganizationalUnitsForParent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListOrganizationalUnitsForParent', listOrganizationalUnitsForParent_nextToken - The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

$sel:maxResults:ListOrganizationalUnitsForParent', listOrganizationalUnitsForParent_maxResults - The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

$sel:parentId:ListOrganizationalUnitsForParent', listOrganizationalUnitsForParent_parentId - The unique identifier (ID) of the root or OU whose child OUs you want to list.

The regex pattern for a parent ID string requires one of the following:

Root - A string that begins with "r-" followed by from 4 to 32 lowercase letters or digits.

Organizational unit (OU) - A string that begins with "ou-" followed by from 4 to 32 lowercase letters or digits (the ID of the root that the OU is in). This string is followed by a second "-" dash and from 8 to 32 additional lowercase letters or digits.

Request Lenses

listOrganizationalUnitsForParent_nextToken :: Lens' ListOrganizationalUnitsForParent (Maybe Text)

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

listOrganizationalUnitsForParent_maxResults :: Lens' ListOrganizationalUnitsForParent (Maybe Natural)

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

listOrganizationalUnitsForParent_parentId :: Lens' ListOrganizationalUnitsForParent Text

The unique identifier (ID) of the root or OU whose child OUs you want to list.

The regex pattern for a parent ID string requires one of the following:

Root - A string that begins with "r-" followed by from 4 to 32 lowercase letters or digits.

Organizational unit (OU) - A string that begins with "ou-" followed by from 4 to 32 lowercase letters or digits (the ID of the root that the OU is in). This string is followed by a second "-" dash and from 8 to 32 additional lowercase letters or digits.

Destructuring the Response

data ListOrganizationalUnitsForParentResponse

See: newListOrganizationalUnitsForParentResponse smart constructor.

Constructors

ListOrganizationalUnitsForParentResponse'

Fields

nextToken :: Maybe Text

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

organizationalUnits :: Maybe [OrganizationalUnit]

A list of the OUs in the specified root or parent OU.

httpStatus :: Int

The response's http status code.

newListOrganizationalUnitsForParentResponse

Arguments

:: Int
$sel:httpStatus:ListOrganizationalUnitsForParentResponse'

-> ListOrganizationalUnitsForParentResponse

Create a value of ListOrganizationalUnitsForParentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListOrganizationalUnitsForParent', listOrganizationalUnitsForParentResponse_nextToken - If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

$sel:organizationalUnits:ListOrganizationalUnitsForParentResponse', listOrganizationalUnitsForParentResponse_organizationalUnits - A list of the OUs in the specified root or parent OU.

$sel:httpStatus:ListOrganizationalUnitsForParentResponse', listOrganizationalUnitsForParentResponse_httpStatus - The response's http status code.

Response Lenses

listOrganizationalUnitsForParentResponse_nextToken :: Lens' ListOrganizationalUnitsForParentResponse (Maybe Text)

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

listOrganizationalUnitsForParentResponse_organizationalUnits :: Lens' ListOrganizationalUnitsForParentResponse (Maybe [OrganizationalUnit])

A list of the OUs in the specified root or parent OU.

listOrganizationalUnitsForParentResponse_httpStatus :: Lens' ListOrganizationalUnitsForParentResponse Int

The response's http status code.