Amazonka.Organizations.Types.Account

Description

Documentation

data Account

Contains information about an AWS account that is a member of an organization.

See: newAccount smart constructor.

Constructors

Account'

Fields

status :: Maybe AccountStatus

The status of the account in the organization.

joinedMethod :: Maybe AccountJoinedMethod

The method by which the account joined the organization.

email :: Maybe (Sensitive Text)

The email address associated with the AWS account.

The regex pattern for this parameter is a string of characters that represents a standard internet email address.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the account.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

joinedTimestamp :: Maybe POSIX

The date the account became a part of the organization.

name :: Maybe (Sensitive Text)

The friendly name of the account.

The regex pattern that is used to validate this parameter is a string of any of the characters in the ASCII character range.

id :: Maybe Text

The unique identifier (ID) of the account.

The regex pattern for an account ID string requires exactly 12 digits.

newAccount :: Account

Create a value of Account with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Account', account_status - The status of the account in the organization.

$sel:joinedMethod:Account', account_joinedMethod - The method by which the account joined the organization.

$sel:email:Account', account_email - The email address associated with the AWS account.

The regex pattern for this parameter is a string of characters that represents a standard internet email address.

$sel:arn:Account', account_arn - The Amazon Resource Name (ARN) of the account.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

$sel:joinedTimestamp:Account', account_joinedTimestamp - The date the account became a part of the organization.

$sel:name:Account', account_name - The friendly name of the account.

The regex pattern that is used to validate this parameter is a string of any of the characters in the ASCII character range.

$sel:id:Account', account_id - The unique identifier (ID) of the account.

The regex pattern for an account ID string requires exactly 12 digits.

account_status :: Lens' Account (Maybe AccountStatus)

The status of the account in the organization.

account_joinedMethod :: Lens' Account (Maybe AccountJoinedMethod)

The method by which the account joined the organization.

account_email :: Lens' Account (Maybe Text)

The email address associated with the AWS account.

The regex pattern for this parameter is a string of characters that represents a standard internet email address.

account_arn :: Lens' Account (Maybe Text)

The Amazon Resource Name (ARN) of the account.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

account_joinedTimestamp :: Lens' Account (Maybe UTCTime)

The date the account became a part of the organization.

account_name :: Lens' Account (Maybe Text)

The friendly name of the account.

The regex pattern that is used to validate this parameter is a string of any of the characters in the ASCII character range.

account_id :: Lens' Account (Maybe Text)

The unique identifier (ID) of the account.

The regex pattern for an account ID string requires exactly 12 digits.