Amazonka.Organizations.Types.CreateAccountStatus

Description

Documentation

data CreateAccountStatus

Contains the status about a CreateAccount or CreateGovCloudAccount request to create an AWS account or an AWS GovCloud (US) account in an organization.

See: newCreateAccountStatus smart constructor.

Constructors

CreateAccountStatus'

Fields

failureReason :: Maybe CreateAccountFailureReason

If the request failed, a description of the reason for the failure.

ACCOUNT_LIMIT_EXCEEDED: The account couldn't be created because you reached the limit on the number of accounts in your organization.

CONCURRENT_ACCOUNT_MODIFICATION: You already submitted a request with the same information.

EMAIL_ALREADY_EXISTS: The account could not be created because another AWS account with that email address already exists.

FAILED_BUSINESS_VALIDATION: The AWS account that owns your organization failed to receive business license validation.

GOVCLOUD_ACCOUNT_ALREADY_EXISTS: The account in the AWS GovCloud (US) Region could not be created because this Region already includes an account with that email address.

IDENTITY_INVALID_BUSINESS_VALIDATION: The AWS account that owns your organization can't complete business license validation because it doesn't have valid identity data.

INVALID_ADDRESS: The account could not be created because the address you provided is not valid.

INVALID_EMAIL: The account could not be created because the email address you provided is not valid.

INTERNAL_FAILURE: The account could not be created because of an internal failure. Try again later. If the problem persists, contact AWS Customer Support.

MISSING_BUSINESS_VALIDATION: The AWS account that owns your organization has not received Business Validation.

MISSING_PAYMENT_INSTRUMENT: You must configure the management account with a valid payment method, such as a credit card.

PENDING_BUSINESS_VALIDATION: The AWS account that owns your organization is still in the process of completing business license validation.

UNKNOWN_BUSINESS_VALIDATION: The AWS account that owns your organization has an unknown issue with business license validation.

state :: Maybe CreateAccountState

The status of the asynchronous request to create an AWS account.

completedTimestamp :: Maybe POSIX

The date and time that the account was created and the request completed.

accountName :: Maybe (Sensitive Text)

The account name given to the account when it was created.

accountId :: Maybe Text

If the account was created successfully, the unique identifier (ID) of the new account.

The regex pattern for an account ID string requires exactly 12 digits.

id :: Maybe Text

The unique identifier (ID) that references this request. You get this value from the response of the initial CreateAccount request to create the account.

The regex pattern for a create account request ID string requires "car-" followed by from 8 to 32 lowercase letters or digits.

govCloudAccountId :: Maybe Text

If the account was created successfully, the unique identifier (ID) of the new account in the AWS GovCloud (US) Region.

requestedTimestamp :: Maybe POSIX

The date and time that the request was made for the account creation.

newCreateAccountStatus :: CreateAccountStatus

Create a value of CreateAccountStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:CreateAccountStatus', createAccountStatus_failureReason - If the request failed, a description of the reason for the failure.

ACCOUNT_LIMIT_EXCEEDED: The account couldn't be created because you reached the limit on the number of accounts in your organization.

CONCURRENT_ACCOUNT_MODIFICATION: You already submitted a request with the same information.

EMAIL_ALREADY_EXISTS: The account could not be created because another AWS account with that email address already exists.

FAILED_BUSINESS_VALIDATION: The AWS account that owns your organization failed to receive business license validation.

GOVCLOUD_ACCOUNT_ALREADY_EXISTS: The account in the AWS GovCloud (US) Region could not be created because this Region already includes an account with that email address.

IDENTITY_INVALID_BUSINESS_VALIDATION: The AWS account that owns your organization can't complete business license validation because it doesn't have valid identity data.

INVALID_ADDRESS: The account could not be created because the address you provided is not valid.

INVALID_EMAIL: The account could not be created because the email address you provided is not valid.

INTERNAL_FAILURE: The account could not be created because of an internal failure. Try again later. If the problem persists, contact AWS Customer Support.

MISSING_BUSINESS_VALIDATION: The AWS account that owns your organization has not received Business Validation.

MISSING_PAYMENT_INSTRUMENT: You must configure the management account with a valid payment method, such as a credit card.

PENDING_BUSINESS_VALIDATION: The AWS account that owns your organization is still in the process of completing business license validation.

UNKNOWN_BUSINESS_VALIDATION: The AWS account that owns your organization has an unknown issue with business license validation.

$sel:state:CreateAccountStatus', createAccountStatus_state - The status of the asynchronous request to create an AWS account.

$sel:completedTimestamp:CreateAccountStatus', createAccountStatus_completedTimestamp - The date and time that the account was created and the request completed.

$sel:accountName:CreateAccountStatus', createAccountStatus_accountName - The account name given to the account when it was created.

$sel:accountId:CreateAccountStatus', createAccountStatus_accountId - If the account was created successfully, the unique identifier (ID) of the new account.

The regex pattern for an account ID string requires exactly 12 digits.

$sel:id:CreateAccountStatus', createAccountStatus_id - The unique identifier (ID) that references this request. You get this value from the response of the initial CreateAccount request to create the account.

The regex pattern for a create account request ID string requires "car-" followed by from 8 to 32 lowercase letters or digits.

$sel:govCloudAccountId:CreateAccountStatus', createAccountStatus_govCloudAccountId - If the account was created successfully, the unique identifier (ID) of the new account in the AWS GovCloud (US) Region.

$sel:requestedTimestamp:CreateAccountStatus', createAccountStatus_requestedTimestamp - The date and time that the request was made for the account creation.

createAccountStatus_failureReason :: Lens' CreateAccountStatus (Maybe CreateAccountFailureReason)

If the request failed, a description of the reason for the failure.

ACCOUNT_LIMIT_EXCEEDED: The account couldn't be created because you reached the limit on the number of accounts in your organization.

CONCURRENT_ACCOUNT_MODIFICATION: You already submitted a request with the same information.

EMAIL_ALREADY_EXISTS: The account could not be created because another AWS account with that email address already exists.

FAILED_BUSINESS_VALIDATION: The AWS account that owns your organization failed to receive business license validation.

GOVCLOUD_ACCOUNT_ALREADY_EXISTS: The account in the AWS GovCloud (US) Region could not be created because this Region already includes an account with that email address.

IDENTITY_INVALID_BUSINESS_VALIDATION: The AWS account that owns your organization can't complete business license validation because it doesn't have valid identity data.

INVALID_ADDRESS: The account could not be created because the address you provided is not valid.

INVALID_EMAIL: The account could not be created because the email address you provided is not valid.

INTERNAL_FAILURE: The account could not be created because of an internal failure. Try again later. If the problem persists, contact AWS Customer Support.

MISSING_BUSINESS_VALIDATION: The AWS account that owns your organization has not received Business Validation.

MISSING_PAYMENT_INSTRUMENT: You must configure the management account with a valid payment method, such as a credit card.

PENDING_BUSINESS_VALIDATION: The AWS account that owns your organization is still in the process of completing business license validation.

UNKNOWN_BUSINESS_VALIDATION: The AWS account that owns your organization has an unknown issue with business license validation.

createAccountStatus_state :: Lens' CreateAccountStatus (Maybe CreateAccountState)

The status of the asynchronous request to create an AWS account.

createAccountStatus_completedTimestamp :: Lens' CreateAccountStatus (Maybe UTCTime)

The date and time that the account was created and the request completed.

createAccountStatus_accountName :: Lens' CreateAccountStatus (Maybe Text)

The account name given to the account when it was created.

createAccountStatus_accountId :: Lens' CreateAccountStatus (Maybe Text)

If the account was created successfully, the unique identifier (ID) of the new account.

The regex pattern for an account ID string requires exactly 12 digits.

createAccountStatus_id :: Lens' CreateAccountStatus (Maybe Text)

The unique identifier (ID) that references this request. You get this value from the response of the initial CreateAccount request to create the account.

The regex pattern for a create account request ID string requires "car-" followed by from 8 to 32 lowercase letters or digits.

createAccountStatus_govCloudAccountId :: Lens' CreateAccountStatus (Maybe Text)

If the account was created successfully, the unique identifier (ID) of the new account in the AWS GovCloud (US) Region.

createAccountStatus_requestedTimestamp :: Lens' CreateAccountStatus (Maybe UTCTime)

The date and time that the request was made for the account creation.