Amazonka.Organizations.Types.HandshakeResource

Description

Documentation

data HandshakeResource

Contains additional data that is needed to process a handshake.

See: newHandshakeResource smart constructor.

Constructors

HandshakeResource'

Fields

value :: Maybe (Sensitive Text)

The information that is passed to the other party in the handshake. The format of the value string must match the requirements of the specified type.

resources :: Maybe [HandshakeResource]

When needed, contains an additional array of HandshakeResource objects.

type' :: Maybe HandshakeResourceType

The type of information being passed, specifying how the value is to be interpreted by the other party:

ACCOUNT - Specifies an AWS account ID number.

ORGANIZATION - Specifies an organization ID number.

EMAIL - Specifies the email address that is associated with the account that receives the handshake.

OWNER_EMAIL - Specifies the email address associated with the management account. Included as information about an organization.

OWNER_NAME - Specifies the name associated with the management account. Included as information about an organization.

NOTES - Additional text provided by the handshake initiator and intended for the recipient to read.

newHandshakeResource :: HandshakeResource

Create a value of HandshakeResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:HandshakeResource', handshakeResource_value - The information that is passed to the other party in the handshake. The format of the value string must match the requirements of the specified type.

$sel:resources:HandshakeResource', handshakeResource_resources - When needed, contains an additional array of HandshakeResource objects.

$sel:type':HandshakeResource', handshakeResource_type - The type of information being passed, specifying how the value is to be interpreted by the other party:

ACCOUNT - Specifies an AWS account ID number.

ORGANIZATION - Specifies an organization ID number.

EMAIL - Specifies the email address that is associated with the account that receives the handshake.

OWNER_EMAIL - Specifies the email address associated with the management account. Included as information about an organization.

OWNER_NAME - Specifies the name associated with the management account. Included as information about an organization.

NOTES - Additional text provided by the handshake initiator and intended for the recipient to read.

handshakeResource_value :: Lens' HandshakeResource (Maybe Text)

The information that is passed to the other party in the handshake. The format of the value string must match the requirements of the specified type.

handshakeResource_resources :: Lens' HandshakeResource (Maybe [HandshakeResource])

When needed, contains an additional array of HandshakeResource objects.

handshakeResource_type :: Lens' HandshakeResource (Maybe HandshakeResourceType)

The type of information being passed, specifying how the value is to be interpreted by the other party:

ACCOUNT - Specifies an AWS account ID number.

ORGANIZATION - Specifies an organization ID number.

EMAIL - Specifies the email address that is associated with the account that receives the handshake.

OWNER_EMAIL - Specifies the email address associated with the management account. Included as information about an organization.

OWNER_NAME - Specifies the name associated with the management account. Included as information about an organization.

NOTES - Additional text provided by the handshake initiator and intended for the recipient to read.