Amazonka.Outposts

Description

Derived from API version 2019-12-03 of the AWS service descriptions, licensed under Apache 2.0.

AWS Outposts is a fully managed service that extends AWS infrastructure, APIs, and tools to customer premises. By providing local access to AWS managed infrastructure, AWS Outposts enables customers to build and run applications on premises using the same programming interfaces as in AWS Regions, while using local compute and storage resources for lower latency and local data processing needs.

Service Configuration

defaultService :: Service

API version 2019-12-03 of the Amazon Outposts SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Outposts.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

A parameter is not valid.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have permission to perform this operation.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Updating or deleting this resource can cause an inconsistent state.

NotFoundException

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified request is not valid.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded a service quota.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error has occurred.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DeleteOutpost

data DeleteOutpost

See: newDeleteOutpost smart constructor.

Constructors

DeleteOutpost' Text

newDeleteOutpost

Arguments

:: Text
$sel:outpostId:DeleteOutpost'

-> DeleteOutpost

Create a value of DeleteOutpost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outpostId:DeleteOutpost', deleteOutpost_outpostId - The ID of the Outpost.

data DeleteOutpostResponse

See: newDeleteOutpostResponse smart constructor.

Constructors

DeleteOutpostResponse' Int

newDeleteOutpostResponse

Arguments

:: Int
$sel:httpStatus:DeleteOutpostResponse'

-> DeleteOutpostResponse

Create a value of DeleteOutpostResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteOutpostResponse', deleteOutpostResponse_httpStatus - The response's http status code.

DeleteSite

data DeleteSite

See: newDeleteSite smart constructor.

Constructors

DeleteSite' Text

newDeleteSite

Arguments

:: Text
$sel:siteId:DeleteSite'

-> DeleteSite

Create a value of DeleteSite with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:siteId:DeleteSite', deleteSite_siteId - Undocumented member.

data DeleteSiteResponse

See: newDeleteSiteResponse smart constructor.

Constructors

DeleteSiteResponse' Int

newDeleteSiteResponse

Arguments

:: Int
$sel:httpStatus:DeleteSiteResponse'

-> DeleteSiteResponse

Create a value of DeleteSiteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteSiteResponse', deleteSiteResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The resource tags.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

ListOutposts

data ListOutposts

See: newListOutposts smart constructor.

Constructors

ListOutposts' (Maybe (NonEmpty Text)) (Maybe (NonEmpty Text)) (Maybe Text) (Maybe (NonEmpty Text)) (Maybe Natural)

newListOutposts :: ListOutposts

Create a value of ListOutposts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZoneFilter:ListOutposts', listOutposts_availabilityZoneFilter - A filter for the Availibility Zone (us-east-1a) of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

$sel:lifeCycleStatusFilter:ListOutposts', listOutposts_lifeCycleStatusFilter - A filter for the lifecycle status of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

$sel:nextToken:ListOutposts', listOutposts_nextToken - Undocumented member.

$sel:availabilityZoneIdFilter:ListOutposts', listOutposts_availabilityZoneIdFilter - A filter for the AZ IDs (use1-az1) of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

$sel:maxResults:ListOutposts', listOutposts_maxResults - Undocumented member.

data ListOutpostsResponse

See: newListOutpostsResponse smart constructor.

Constructors

ListOutpostsResponse' (Maybe Text) (Maybe [Outpost]) Int

newListOutpostsResponse

Arguments

:: Int
$sel:httpStatus:ListOutpostsResponse'

-> ListOutpostsResponse

Create a value of ListOutpostsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListOutposts', listOutpostsResponse_nextToken - Undocumented member.

$sel:outposts:ListOutpostsResponse', listOutpostsResponse_outposts - Undocumented member.

$sel:httpStatus:ListOutpostsResponse', listOutpostsResponse_httpStatus - The response's http status code.

ListSites

data ListSites

See: newListSites smart constructor.

Constructors

ListSites' (Maybe Text) (Maybe Natural)

newListSites :: ListSites

Create a value of ListSites with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSites', listSites_nextToken - Undocumented member.

$sel:maxResults:ListSites', listSites_maxResults - Undocumented member.

data ListSitesResponse

See: newListSitesResponse smart constructor.

Constructors

ListSitesResponse' (Maybe Text) (Maybe [Site]) Int

newListSitesResponse

Arguments

:: Int
$sel:httpStatus:ListSitesResponse'

-> ListSitesResponse

Create a value of ListSitesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSites', listSitesResponse_nextToken - Undocumented member.

$sel:sites:ListSitesResponse', listSitesResponse_sites - Undocumented member.

$sel:httpStatus:ListSitesResponse', listSitesResponse_httpStatus - The response's http status code.

CreateOrder

data CreateOrder

See: newCreateOrder smart constructor.

Constructors

CreateOrder' (Maybe PaymentTerm) Text (NonEmpty LineItemRequest) PaymentOption

newCreateOrder

Arguments

:: Text
$sel:outpostIdentifier:CreateOrder'

-> NonEmpty LineItemRequest
$sel:lineItems:CreateOrder'

-> PaymentOption
$sel:paymentOption:CreateOrder'

-> CreateOrder

Create a value of CreateOrder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paymentTerm:CreateOrder', createOrder_paymentTerm - The payment terms for the order.

$sel:outpostIdentifier:CreateOrder', createOrder_outpostIdentifier - The ID or the Amazon Resource Name (ARN) of the Outpost.

$sel:lineItems:CreateOrder', createOrder_lineItems - The line items that make up the order.

$sel:paymentOption:CreateOrder', createOrder_paymentOption - The payment option for the order.

data CreateOrderResponse

See: newCreateOrderResponse smart constructor.

Constructors

CreateOrderResponse' (Maybe Order) Int

newCreateOrderResponse

Arguments

:: Int
$sel:httpStatus:CreateOrderResponse'

-> CreateOrderResponse

Create a value of CreateOrderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:order:CreateOrderResponse', createOrderResponse_order - Information about this order.

$sel:httpStatus:CreateOrderResponse', createOrderResponse_httpStatus - The response's http status code.

GetOutpostInstanceTypes

data GetOutpostInstanceTypes

See: newGetOutpostInstanceTypes smart constructor.

Constructors

GetOutpostInstanceTypes' (Maybe Text) (Maybe Natural) Text

newGetOutpostInstanceTypes

Arguments

:: Text
$sel:outpostId:GetOutpostInstanceTypes'

-> GetOutpostInstanceTypes

Create a value of GetOutpostInstanceTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetOutpostInstanceTypes', getOutpostInstanceTypes_nextToken - Undocumented member.

$sel:maxResults:GetOutpostInstanceTypes', getOutpostInstanceTypes_maxResults - Undocumented member.

$sel:outpostId:GetOutpostInstanceTypes', getOutpostInstanceTypes_outpostId - The ID of the Outpost.

data GetOutpostInstanceTypesResponse

See: newGetOutpostInstanceTypesResponse smart constructor.

Constructors

GetOutpostInstanceTypesResponse' (Maybe [InstanceTypeItem]) (Maybe Text) (Maybe Text) (Maybe Text) Int

newGetOutpostInstanceTypesResponse

Arguments

:: Int
$sel:httpStatus:GetOutpostInstanceTypesResponse'

-> GetOutpostInstanceTypesResponse

Create a value of GetOutpostInstanceTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceTypes:GetOutpostInstanceTypesResponse', getOutpostInstanceTypesResponse_instanceTypes - Undocumented member.

$sel:outpostArn:GetOutpostInstanceTypesResponse', getOutpostInstanceTypesResponse_outpostArn - Undocumented member.

$sel:nextToken:GetOutpostInstanceTypes', getOutpostInstanceTypesResponse_nextToken - Undocumented member.

$sel:outpostId:GetOutpostInstanceTypes', getOutpostInstanceTypesResponse_outpostId - The ID of the Outpost.

$sel:httpStatus:GetOutpostInstanceTypesResponse', getOutpostInstanceTypesResponse_httpStatus - The response's http status code.

CreateOutpost

data CreateOutpost

See: newCreateOutpost smart constructor.

Constructors

CreateOutpost' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Text Text

newCreateOutpost

Arguments

:: Text
$sel:name:CreateOutpost'

-> Text
$sel:siteId:CreateOutpost'

-> CreateOutpost

Create a value of CreateOutpost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZoneId:CreateOutpost', createOutpost_availabilityZoneId - Undocumented member.

$sel:availabilityZone:CreateOutpost', createOutpost_availabilityZone - Undocumented member.

$sel:description:CreateOutpost', createOutpost_description - Undocumented member.

$sel:tags:CreateOutpost', createOutpost_tags - The tags to apply to the Outpost.

$sel:name:CreateOutpost', createOutpost_name - Undocumented member.

$sel:siteId:CreateOutpost', createOutpost_siteId - Undocumented member.

data CreateOutpostResponse

See: newCreateOutpostResponse smart constructor.

Constructors

CreateOutpostResponse' (Maybe Outpost) Int

newCreateOutpostResponse

Arguments

:: Int
$sel:httpStatus:CreateOutpostResponse'

-> CreateOutpostResponse

Create a value of CreateOutpostResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outpost:CreateOutpostResponse', createOutpostResponse_outpost - Undocumented member.

$sel:httpStatus:CreateOutpostResponse', createOutpostResponse_httpStatus - The response's http status code.

GetOutpost

data GetOutpost

See: newGetOutpost smart constructor.

Constructors

GetOutpost' Text

newGetOutpost

Arguments

:: Text
$sel:outpostId:GetOutpost'

-> GetOutpost

Create a value of GetOutpost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outpostId:GetOutpost', getOutpost_outpostId - The ID of the Outpost.

data GetOutpostResponse

See: newGetOutpostResponse smart constructor.

Constructors

GetOutpostResponse' (Maybe Outpost) Int

newGetOutpostResponse

Arguments

:: Int
$sel:httpStatus:GetOutpostResponse'

-> GetOutpostResponse

Create a value of GetOutpostResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outpost:GetOutpostResponse', getOutpostResponse_outpost - Undocumented member.

$sel:httpStatus:GetOutpostResponse', getOutpostResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:tags:TagResource', tagResource_tags - The tags to add to the resource.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text (NonEmpty Text)

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> NonEmpty Text
$sel:tagKeys:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The tag keys.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

Types

OrderStatus

newtype OrderStatus

Constructors

OrderStatus'

Fields

fromOrderStatus :: Text

Bundled Patterns

pattern OrderStatus_CANCELLED :: OrderStatus

pattern OrderStatus_FULFILLED :: OrderStatus

pattern OrderStatus_INSTALLING :: OrderStatus

pattern OrderStatus_PENDING :: OrderStatus

pattern OrderStatus_PROCESSING :: OrderStatus

pattern OrderStatus_RECEIVED :: OrderStatus

PaymentOption

newtype PaymentOption

Constructors

PaymentOption'

Fields

fromPaymentOption :: Text

Bundled Patterns

pattern PaymentOption_ALL_UPFRONT :: PaymentOption

pattern PaymentOption_NO_UPFRONT :: PaymentOption

pattern PaymentOption_PARTIAL_UPFRONT :: PaymentOption

PaymentTerm

newtype PaymentTerm

Constructors

PaymentTerm'

Fields

fromPaymentTerm :: Text

Bundled Patterns

pattern PaymentTerm_THREE_YEARS :: PaymentTerm

InstanceTypeItem

data InstanceTypeItem

Information about an instance type.

See: newInstanceTypeItem smart constructor.

Constructors

InstanceTypeItem' (Maybe Text)

newInstanceTypeItem :: InstanceTypeItem

Create a value of InstanceTypeItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceType:InstanceTypeItem', instanceTypeItem_instanceType - Undocumented member.

LineItem

data LineItem

Information about a line item.

See: newLineItem smart constructor.

Constructors

LineItem' (Maybe Text) (Maybe Natural) (Maybe Text) (Maybe Text)

newLineItem :: LineItem

Create a value of LineItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:LineItem', lineItem_status - The status of the line item.

$sel:quantity:LineItem', lineItem_quantity - The quantity of the line item.

$sel:catalogItemId:LineItem', lineItem_catalogItemId - The ID of the catalog item.

$sel:lineItemId:LineItem', lineItem_lineItemId - The ID of the line item.

LineItemRequest

data LineItemRequest

Information about a line item request.

See: newLineItemRequest smart constructor.

Constructors

LineItemRequest' (Maybe Natural) (Maybe Text)

newLineItemRequest :: LineItemRequest

Create a value of LineItemRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quantity:LineItemRequest', lineItemRequest_quantity - The quantity of a line item request.

$sel:catalogItemId:LineItemRequest', lineItemRequest_catalogItemId - The ID of the catalog item.

Order

data Order

Information about an order.

See: newOrder smart constructor.

Constructors

Order' (Maybe OrderStatus) (Maybe POSIX) (Maybe [LineItem]) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe PaymentOption)

newOrder :: Order

Create a value of Order with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Order', order_status - The status of the order

$sel:orderSubmissionDate:Order', order_orderSubmissionDate - The submission date for the order.

$sel:lineItems:Order', order_lineItems - The line items for the order

$sel:orderFulfilledDate:Order', order_orderFulfilledDate - The fulfillment date of the order.

$sel:orderId:Order', order_orderId - The ID of the order.

$sel:outpostId:Order', order_outpostId - The ID of the Outpost.

$sel:paymentOption:Order', order_paymentOption - The payment option for the order.

Outpost

data Outpost

Information about an Outpost.

See: newOutpost smart constructor.

Constructors

Outpost' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text))

newOutpost :: Outpost

Create a value of Outpost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZoneId:Outpost', outpost_availabilityZoneId - Undocumented member.

$sel:outpostArn:Outpost', outpost_outpostArn - Undocumented member.

$sel:ownerId:Outpost', outpost_ownerId - Undocumented member.

$sel:availabilityZone:Outpost', outpost_availabilityZone - Undocumented member.

$sel:name:Outpost', outpost_name - Undocumented member.

$sel:lifeCycleStatus:Outpost', outpost_lifeCycleStatus - Undocumented member.

$sel:outpostId:Outpost', outpost_outpostId - The ID of the Outpost.

$sel:siteId:Outpost', outpost_siteId - Undocumented member.

$sel:siteArn:Outpost', outpost_siteArn - Undocumented member.

$sel:description:Outpost', outpost_description - Undocumented member.

$sel:tags:Outpost', outpost_tags - The Outpost tags.

Site

data Site

Information about a site.

See: newSite smart constructor.

Constructors

Site' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text))

newSite :: Site

Create a value of Site with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:Site', site_accountId - Undocumented member.

$sel:name:Site', site_name - Undocumented member.

$sel:siteId:Site', site_siteId - Undocumented member.

$sel:siteArn:Site', site_siteArn - Undocumented member.

$sel:description:Site', site_description - Undocumented member.

$sel:tags:Site', site_tags - The site tags.