Amazonka.Outposts.ListOutposts

Description

Create a list of the Outposts for your AWS account. Add filters to your request to return a more specific list of results. Use filters to match an Outpost lifecycle status, Availibility Zone (us-east-1a), and AZ ID (use1-az1).

If you specify multiple filters, the filters are joined with an AND, and the request returns only results that match all of the specified filters.

Creating a Request

data ListOutposts

See: newListOutposts smart constructor.

Constructors

ListOutposts'

Fields

availabilityZoneFilter :: Maybe (NonEmpty Text)

A filter for the Availibility Zone (us-east-1a) of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

lifeCycleStatusFilter :: Maybe (NonEmpty Text)

A filter for the lifecycle status of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

nextToken :: Maybe Text

availabilityZoneIdFilter :: Maybe (NonEmpty Text)

A filter for the AZ IDs (use1-az1) of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

maxResults :: Maybe Natural

newListOutposts :: ListOutposts

Create a value of ListOutposts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZoneFilter:ListOutposts', listOutposts_availabilityZoneFilter - A filter for the Availibility Zone (us-east-1a) of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

$sel:lifeCycleStatusFilter:ListOutposts', listOutposts_lifeCycleStatusFilter - A filter for the lifecycle status of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

$sel:nextToken:ListOutposts', listOutposts_nextToken - Undocumented member.

$sel:availabilityZoneIdFilter:ListOutposts', listOutposts_availabilityZoneIdFilter - A filter for the AZ IDs (use1-az1) of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

$sel:maxResults:ListOutposts', listOutposts_maxResults - Undocumented member.

Request Lenses

listOutposts_availabilityZoneFilter :: Lens' ListOutposts (Maybe (NonEmpty Text))

A filter for the Availibility Zone (us-east-1a) of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

listOutposts_lifeCycleStatusFilter :: Lens' ListOutposts (Maybe (NonEmpty Text))

A filter for the lifecycle status of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

listOutposts_nextToken :: Lens' ListOutposts (Maybe Text)

Undocumented member.

listOutposts_availabilityZoneIdFilter :: Lens' ListOutposts (Maybe (NonEmpty Text))

A filter for the AZ IDs (use1-az1) of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

listOutposts_maxResults :: Lens' ListOutposts (Maybe Natural)

Undocumented member.

Destructuring the Response

data ListOutpostsResponse

See: newListOutpostsResponse smart constructor.

Constructors

ListOutpostsResponse'

Fields

nextToken :: Maybe Text

outposts :: Maybe [Outpost]

httpStatus :: Int

The response's http status code.

newListOutpostsResponse

Arguments

:: Int
$sel:httpStatus:ListOutpostsResponse'

-> ListOutpostsResponse

Create a value of ListOutpostsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListOutposts', listOutpostsResponse_nextToken - Undocumented member.

$sel:outposts:ListOutpostsResponse', listOutpostsResponse_outposts - Undocumented member.

$sel:httpStatus:ListOutpostsResponse', listOutpostsResponse_httpStatus - The response's http status code.

Response Lenses

listOutpostsResponse_nextToken :: Lens' ListOutpostsResponse (Maybe Text)

Undocumented member.

listOutpostsResponse_outposts :: Lens' ListOutpostsResponse (Maybe [Outpost])

Undocumented member.

listOutpostsResponse_httpStatus :: Lens' ListOutpostsResponse Int

The response's http status code.