Amazonka.Outposts.Types.Order

Description

Documentation

data Order

Information about an order.

See: newOrder smart constructor.

Constructors

Order'

Fields

status :: Maybe OrderStatus

The status of the order

orderSubmissionDate :: Maybe POSIX

The submission date for the order.

lineItems :: Maybe [LineItem]

The line items for the order

orderFulfilledDate :: Maybe POSIX

The fulfillment date of the order.

orderId :: Maybe Text

The ID of the order.

outpostId :: Maybe Text

The ID of the Outpost.

paymentOption :: Maybe PaymentOption

The payment option for the order.

newOrder :: Order

Create a value of Order with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Order', order_status - The status of the order

$sel:orderSubmissionDate:Order', order_orderSubmissionDate - The submission date for the order.

$sel:lineItems:Order', order_lineItems - The line items for the order

$sel:orderFulfilledDate:Order', order_orderFulfilledDate - The fulfillment date of the order.

$sel:orderId:Order', order_orderId - The ID of the order.

$sel:outpostId:Order', order_outpostId - The ID of the Outpost.

$sel:paymentOption:Order', order_paymentOption - The payment option for the order.

order_status :: Lens' Order (Maybe OrderStatus)

The status of the order

order_orderSubmissionDate :: Lens' Order (Maybe UTCTime)

The submission date for the order.

order_lineItems :: Lens' Order (Maybe [LineItem])

The line items for the order

order_orderFulfilledDate :: Lens' Order (Maybe UTCTime)

The fulfillment date of the order.

order_orderId :: Lens' Order (Maybe Text)

The ID of the order.

order_outpostId :: Lens' Order (Maybe Text)

The ID of the Outpost.

order_paymentOption :: Lens' Order (Maybe PaymentOption)

The payment option for the order.