Amazonka.Panorama.DescribeApplicationInstance

Description

Returns information about an application instance on a device.

Creating a Request

data DescribeApplicationInstance

See: newDescribeApplicationInstance smart constructor.

Constructors

DescribeApplicationInstance'

Fields

applicationInstanceId :: Text

The application instance's ID.

newDescribeApplicationInstance

Arguments

:: Text
$sel:applicationInstanceId:DescribeApplicationInstance'

-> DescribeApplicationInstance

Create a value of DescribeApplicationInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationInstanceId:DescribeApplicationInstance', describeApplicationInstance_applicationInstanceId - The application instance's ID.

Request Lenses

describeApplicationInstance_applicationInstanceId :: Lens' DescribeApplicationInstance Text

The application instance's ID.

Destructuring the Response

data DescribeApplicationInstanceResponse

See: newDescribeApplicationInstanceResponse smart constructor.

Constructors

DescribeApplicationInstanceResponse'

Fields

status :: Maybe ApplicationInstanceStatus

The application instance's status.

statusDescription :: Maybe Text

The application instance's status description.

lastUpdatedTime :: Maybe POSIX

The application instance was updated.

arn :: Maybe Text

The application instance's ARN.

createdTime :: Maybe POSIX

When the application instance was created.

defaultRuntimeContextDevice :: Maybe Text

The device's ID.

defaultRuntimeContextDeviceName :: Maybe Text

The device's bane.

name :: Maybe Text

The application instance's name.

runtimeRoleArn :: Maybe Text

The application instance's runtime role ARN.

healthStatus :: Maybe ApplicationInstanceHealthStatus

The application instance's health status.

applicationInstanceId :: Maybe Text

The application instance's ID.

description :: Maybe Text

The application instance's description.

tags :: Maybe (HashMap Text Text)

The application instance's tags.

applicationInstanceIdToReplace :: Maybe Text

The ID of the application instance that this instance replaced.

httpStatus :: Int

The response's http status code.

newDescribeApplicationInstanceResponse

Arguments

:: Int
$sel:httpStatus:DescribeApplicationInstanceResponse'

-> DescribeApplicationInstanceResponse

Create a value of DescribeApplicationInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_status - The application instance's status.

$sel:statusDescription:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_statusDescription - The application instance's status description.

$sel:lastUpdatedTime:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_lastUpdatedTime - The application instance was updated.

$sel:arn:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_arn - The application instance's ARN.

$sel:createdTime:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_createdTime - When the application instance was created.

$sel:defaultRuntimeContextDevice:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_defaultRuntimeContextDevice - The device's ID.

$sel:defaultRuntimeContextDeviceName:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_defaultRuntimeContextDeviceName - The device's bane.

$sel:name:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_name - The application instance's name.

$sel:runtimeRoleArn:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_runtimeRoleArn - The application instance's runtime role ARN.

$sel:healthStatus:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_healthStatus - The application instance's health status.

$sel:applicationInstanceId:DescribeApplicationInstance', describeApplicationInstanceResponse_applicationInstanceId - The application instance's ID.

$sel:description:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_description - The application instance's description.

$sel:tags:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_tags - The application instance's tags.

$sel:applicationInstanceIdToReplace:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_applicationInstanceIdToReplace - The ID of the application instance that this instance replaced.

$sel:httpStatus:DescribeApplicationInstanceResponse', describeApplicationInstanceResponse_httpStatus - The response's http status code.

Response Lenses

describeApplicationInstanceResponse_status :: Lens' DescribeApplicationInstanceResponse (Maybe ApplicationInstanceStatus)

The application instance's status.

describeApplicationInstanceResponse_statusDescription :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The application instance's status description.

describeApplicationInstanceResponse_lastUpdatedTime :: Lens' DescribeApplicationInstanceResponse (Maybe UTCTime)

The application instance was updated.

describeApplicationInstanceResponse_arn :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The application instance's ARN.

describeApplicationInstanceResponse_createdTime :: Lens' DescribeApplicationInstanceResponse (Maybe UTCTime)

When the application instance was created.

describeApplicationInstanceResponse_defaultRuntimeContextDevice :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The device's ID.

describeApplicationInstanceResponse_defaultRuntimeContextDeviceName :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The device's bane.

describeApplicationInstanceResponse_name :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The application instance's name.

describeApplicationInstanceResponse_runtimeRoleArn :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The application instance's runtime role ARN.

describeApplicationInstanceResponse_healthStatus :: Lens' DescribeApplicationInstanceResponse (Maybe ApplicationInstanceHealthStatus)

The application instance's health status.

describeApplicationInstanceResponse_applicationInstanceId :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The application instance's ID.

describeApplicationInstanceResponse_description :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The application instance's description.

describeApplicationInstanceResponse_tags :: Lens' DescribeApplicationInstanceResponse (Maybe (HashMap Text Text))

The application instance's tags.

describeApplicationInstanceResponse_applicationInstanceIdToReplace :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The ID of the application instance that this instance replaced.

describeApplicationInstanceResponse_httpStatus :: Lens' DescribeApplicationInstanceResponse Int

The response's http status code.