Amazonka.Panorama.DescribeDevice

Description

Returns information about a device.

Creating a Request

data DescribeDevice

See: newDescribeDevice smart constructor.

Constructors

DescribeDevice'

Fields

deviceId :: Text

The device's ID.

newDescribeDevice

Arguments

:: Text
$sel:deviceId:DescribeDevice'

-> DescribeDevice

Create a value of DescribeDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:DescribeDevice', describeDevice_deviceId - The device's ID.

Request Lenses

describeDevice_deviceId :: Lens' DescribeDevice Text

The device's ID.

Destructuring the Response

data DescribeDeviceResponse

See: newDescribeDeviceResponse smart constructor.

Constructors

DescribeDeviceResponse'

Fields

latestSoftware :: Maybe Text

The latest software version available for the device.

provisioningStatus :: Maybe DeviceStatus

The device's provisioning status.

arn :: Maybe Text

The device's ARN.

createdTime :: Maybe POSIX

When the device was created.

currentSoftware :: Maybe Text

The device's current software version.

name :: Maybe Text

The device's name.

deviceConnectionStatus :: Maybe DeviceConnectionStatus

The device's connection status.

deviceId :: Maybe Text

The device's ID.

type' :: Maybe DeviceType

The device's type.

leaseExpirationTime :: Maybe POSIX

The device's lease expiration time.

serialNumber :: Maybe Text

The device's serial number.

currentNetworkingStatus :: Maybe NetworkStatus

The device's networking status.

description :: Maybe Text

The device's description.

networkingConfiguration :: Maybe NetworkPayload

The device's networking configuration.

tags :: Maybe (HashMap Text Text)

The device's tags.

httpStatus :: Int

The response's http status code.

newDescribeDeviceResponse

Arguments

:: Int
$sel:httpStatus:DescribeDeviceResponse'

-> DescribeDeviceResponse

Create a value of DescribeDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestSoftware:DescribeDeviceResponse', describeDeviceResponse_latestSoftware - The latest software version available for the device.

$sel:provisioningStatus:DescribeDeviceResponse', describeDeviceResponse_provisioningStatus - The device's provisioning status.

$sel:arn:DescribeDeviceResponse', describeDeviceResponse_arn - The device's ARN.

$sel:createdTime:DescribeDeviceResponse', describeDeviceResponse_createdTime - When the device was created.

$sel:currentSoftware:DescribeDeviceResponse', describeDeviceResponse_currentSoftware - The device's current software version.

$sel:name:DescribeDeviceResponse', describeDeviceResponse_name - The device's name.

$sel:deviceConnectionStatus:DescribeDeviceResponse', describeDeviceResponse_deviceConnectionStatus - The device's connection status.

$sel:deviceId:DescribeDevice', describeDeviceResponse_deviceId - The device's ID.

$sel:type':DescribeDeviceResponse', describeDeviceResponse_type - The device's type.

$sel:leaseExpirationTime:DescribeDeviceResponse', describeDeviceResponse_leaseExpirationTime - The device's lease expiration time.

$sel:serialNumber:DescribeDeviceResponse', describeDeviceResponse_serialNumber - The device's serial number.

$sel:currentNetworkingStatus:DescribeDeviceResponse', describeDeviceResponse_currentNetworkingStatus - The device's networking status.

$sel:description:DescribeDeviceResponse', describeDeviceResponse_description - The device's description.

$sel:networkingConfiguration:DescribeDeviceResponse', describeDeviceResponse_networkingConfiguration - The device's networking configuration.

$sel:tags:DescribeDeviceResponse', describeDeviceResponse_tags - The device's tags.

$sel:httpStatus:DescribeDeviceResponse', describeDeviceResponse_httpStatus - The response's http status code.

Response Lenses

describeDeviceResponse_latestSoftware :: Lens' DescribeDeviceResponse (Maybe Text)

The latest software version available for the device.

describeDeviceResponse_provisioningStatus :: Lens' DescribeDeviceResponse (Maybe DeviceStatus)

The device's provisioning status.

describeDeviceResponse_arn :: Lens' DescribeDeviceResponse (Maybe Text)

The device's ARN.

describeDeviceResponse_createdTime :: Lens' DescribeDeviceResponse (Maybe UTCTime)

When the device was created.

describeDeviceResponse_currentSoftware :: Lens' DescribeDeviceResponse (Maybe Text)

The device's current software version.

describeDeviceResponse_name :: Lens' DescribeDeviceResponse (Maybe Text)

The device's name.

describeDeviceResponse_deviceConnectionStatus :: Lens' DescribeDeviceResponse (Maybe DeviceConnectionStatus)

The device's connection status.

describeDeviceResponse_deviceId :: Lens' DescribeDeviceResponse (Maybe Text)

The device's ID.

describeDeviceResponse_type :: Lens' DescribeDeviceResponse (Maybe DeviceType)

The device's type.

describeDeviceResponse_leaseExpirationTime :: Lens' DescribeDeviceResponse (Maybe UTCTime)

The device's lease expiration time.

describeDeviceResponse_serialNumber :: Lens' DescribeDeviceResponse (Maybe Text)

The device's serial number.

describeDeviceResponse_currentNetworkingStatus :: Lens' DescribeDeviceResponse (Maybe NetworkStatus)

The device's networking status.

describeDeviceResponse_description :: Lens' DescribeDeviceResponse (Maybe Text)

The device's description.

describeDeviceResponse_networkingConfiguration :: Lens' DescribeDeviceResponse (Maybe NetworkPayload)

The device's networking configuration.

describeDeviceResponse_tags :: Lens' DescribeDeviceResponse (Maybe (HashMap Text Text))

The device's tags.

describeDeviceResponse_httpStatus :: Lens' DescribeDeviceResponse Int

The response's http status code.