Amazonka.Panorama.DescribeNodeFromTemplateJob

Description

Returns information about a job to create a camera stream node.

Creating a Request

data DescribeNodeFromTemplateJob

See: newDescribeNodeFromTemplateJob smart constructor.

Constructors

DescribeNodeFromTemplateJob'

Fields

jobId :: Text

The job's ID.

newDescribeNodeFromTemplateJob

Arguments

:: Text
$sel:jobId:DescribeNodeFromTemplateJob'

-> DescribeNodeFromTemplateJob

Create a value of DescribeNodeFromTemplateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:DescribeNodeFromTemplateJob', describeNodeFromTemplateJob_jobId - The job's ID.

Request Lenses

describeNodeFromTemplateJob_jobId :: Lens' DescribeNodeFromTemplateJob Text

The job's ID.

Destructuring the Response

data DescribeNodeFromTemplateJobResponse

See: newDescribeNodeFromTemplateJobResponse smart constructor.

Constructors

DescribeNodeFromTemplateJobResponse'

Fields

jobTags :: Maybe [JobResourceTags]

The job's tags.

nodeDescription :: Maybe Text

The node's description.

httpStatus :: Int

The response's http status code.

jobId :: Text

The job's ID.

status :: NodeFromTemplateJobStatus

The job's status.

statusMessage :: Text

The job's status message.

createdTime :: POSIX

When the job was created.

lastUpdatedTime :: POSIX

When the job was updated.

outputPackageName :: Text

The job's output package name.

outputPackageVersion :: Text

The job's output package version.

nodeName :: Text

The node's name.

templateType :: TemplateType

The job's template type.

templateParameters :: HashMap Text (Sensitive Text)

The job's template parameters.

newDescribeNodeFromTemplateJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeNodeFromTemplateJobResponse'

-> Text
$sel:jobId:DescribeNodeFromTemplateJob'

-> NodeFromTemplateJobStatus
$sel:status:DescribeNodeFromTemplateJobResponse'

-> Text
$sel:statusMessage:DescribeNodeFromTemplateJobResponse'

-> UTCTime
$sel:createdTime:DescribeNodeFromTemplateJobResponse'

-> UTCTime
$sel:lastUpdatedTime:DescribeNodeFromTemplateJobResponse'

-> Text
$sel:outputPackageName:DescribeNodeFromTemplateJobResponse'

-> Text
$sel:outputPackageVersion:DescribeNodeFromTemplateJobResponse'

-> Text
$sel:nodeName:DescribeNodeFromTemplateJobResponse'

-> TemplateType
$sel:templateType:DescribeNodeFromTemplateJobResponse'

-> DescribeNodeFromTemplateJobResponse

Create a value of DescribeNodeFromTemplateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobTags:DescribeNodeFromTemplateJobResponse', describeNodeFromTemplateJobResponse_jobTags - The job's tags.

$sel:nodeDescription:DescribeNodeFromTemplateJobResponse', describeNodeFromTemplateJobResponse_nodeDescription - The node's description.

$sel:httpStatus:DescribeNodeFromTemplateJobResponse', describeNodeFromTemplateJobResponse_httpStatus - The response's http status code.

$sel:jobId:DescribeNodeFromTemplateJob', describeNodeFromTemplateJobResponse_jobId - The job's ID.

$sel:status:DescribeNodeFromTemplateJobResponse', describeNodeFromTemplateJobResponse_status - The job's status.

$sel:statusMessage:DescribeNodeFromTemplateJobResponse', describeNodeFromTemplateJobResponse_statusMessage - The job's status message.

$sel:createdTime:DescribeNodeFromTemplateJobResponse', describeNodeFromTemplateJobResponse_createdTime - When the job was created.

$sel:lastUpdatedTime:DescribeNodeFromTemplateJobResponse', describeNodeFromTemplateJobResponse_lastUpdatedTime - When the job was updated.

$sel:outputPackageName:DescribeNodeFromTemplateJobResponse', describeNodeFromTemplateJobResponse_outputPackageName - The job's output package name.

$sel:outputPackageVersion:DescribeNodeFromTemplateJobResponse', describeNodeFromTemplateJobResponse_outputPackageVersion - The job's output package version.

$sel:nodeName:DescribeNodeFromTemplateJobResponse', describeNodeFromTemplateJobResponse_nodeName - The node's name.

$sel:templateType:DescribeNodeFromTemplateJobResponse', describeNodeFromTemplateJobResponse_templateType - The job's template type.

$sel:templateParameters:DescribeNodeFromTemplateJobResponse', describeNodeFromTemplateJobResponse_templateParameters - The job's template parameters.

Response Lenses

describeNodeFromTemplateJobResponse_jobTags :: Lens' DescribeNodeFromTemplateJobResponse (Maybe [JobResourceTags])

The job's tags.

describeNodeFromTemplateJobResponse_nodeDescription :: Lens' DescribeNodeFromTemplateJobResponse (Maybe Text)

The node's description.

describeNodeFromTemplateJobResponse_httpStatus :: Lens' DescribeNodeFromTemplateJobResponse Int

The response's http status code.

describeNodeFromTemplateJobResponse_jobId :: Lens' DescribeNodeFromTemplateJobResponse Text

The job's ID.

describeNodeFromTemplateJobResponse_status :: Lens' DescribeNodeFromTemplateJobResponse NodeFromTemplateJobStatus

The job's status.

describeNodeFromTemplateJobResponse_statusMessage :: Lens' DescribeNodeFromTemplateJobResponse Text

The job's status message.

describeNodeFromTemplateJobResponse_createdTime :: Lens' DescribeNodeFromTemplateJobResponse UTCTime

When the job was created.

describeNodeFromTemplateJobResponse_lastUpdatedTime :: Lens' DescribeNodeFromTemplateJobResponse UTCTime

When the job was updated.

describeNodeFromTemplateJobResponse_outputPackageName :: Lens' DescribeNodeFromTemplateJobResponse Text

The job's output package name.

describeNodeFromTemplateJobResponse_outputPackageVersion :: Lens' DescribeNodeFromTemplateJobResponse Text

The job's output package version.

describeNodeFromTemplateJobResponse_nodeName :: Lens' DescribeNodeFromTemplateJobResponse Text

The node's name.

describeNodeFromTemplateJobResponse_templateType :: Lens' DescribeNodeFromTemplateJobResponse TemplateType

The job's template type.

describeNodeFromTemplateJobResponse_templateParameters :: Lens' DescribeNodeFromTemplateJobResponse (HashMap Text Text)

The job's template parameters.