Amazonka.Panorama.DescribePackage

Description

Returns information about a package.

Creating a Request

data DescribePackage

See: newDescribePackage smart constructor.

Constructors

DescribePackage'

Fields

packageId :: Text

The package's ID.

newDescribePackage

Arguments

:: Text
$sel:packageId:DescribePackage'

-> DescribePackage

Create a value of DescribePackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageId:DescribePackage', describePackage_packageId - The package's ID.

Request Lenses

describePackage_packageId :: Lens' DescribePackage Text

The package's ID.

Destructuring the Response

data DescribePackageResponse

See: newDescribePackageResponse smart constructor.

Constructors

DescribePackageResponse'

Fields

writeAccessPrincipalArns :: Maybe [Text]

ARNs of accounts that have write access to the package.

readAccessPrincipalArns :: Maybe [Text]

ARNs of accounts that have read access to the package.

httpStatus :: Int

The response's http status code.

packageId :: Text

The package's ID.

packageName :: Text

The package's name.

arn :: Text

The package's ARN.

storageLocation :: StorageLocation

The package's storage location.

createdTime :: POSIX

When the package was created.

tags :: HashMap Text Text

The package's tags.

newDescribePackageResponse

Arguments

:: Int
$sel:httpStatus:DescribePackageResponse'

-> Text
$sel:packageId:DescribePackage'

-> Text
$sel:packageName:DescribePackageResponse'

-> Text
$sel:arn:DescribePackageResponse'

-> StorageLocation
$sel:storageLocation:DescribePackageResponse'

-> UTCTime
$sel:createdTime:DescribePackageResponse'

-> DescribePackageResponse

Create a value of DescribePackageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:writeAccessPrincipalArns:DescribePackageResponse', describePackageResponse_writeAccessPrincipalArns - ARNs of accounts that have write access to the package.

$sel:readAccessPrincipalArns:DescribePackageResponse', describePackageResponse_readAccessPrincipalArns - ARNs of accounts that have read access to the package.

$sel:httpStatus:DescribePackageResponse', describePackageResponse_httpStatus - The response's http status code.

$sel:packageId:DescribePackage', describePackageResponse_packageId - The package's ID.

$sel:packageName:DescribePackageResponse', describePackageResponse_packageName - The package's name.

$sel:arn:DescribePackageResponse', describePackageResponse_arn - The package's ARN.

$sel:storageLocation:DescribePackageResponse', describePackageResponse_storageLocation - The package's storage location.

$sel:createdTime:DescribePackageResponse', describePackageResponse_createdTime - When the package was created.

$sel:tags:DescribePackageResponse', describePackageResponse_tags - The package's tags.

Response Lenses

describePackageResponse_writeAccessPrincipalArns :: Lens' DescribePackageResponse (Maybe [Text])

ARNs of accounts that have write access to the package.

describePackageResponse_readAccessPrincipalArns :: Lens' DescribePackageResponse (Maybe [Text])

ARNs of accounts that have read access to the package.

describePackageResponse_httpStatus :: Lens' DescribePackageResponse Int

The response's http status code.

describePackageResponse_packageId :: Lens' DescribePackageResponse Text

The package's ID.

describePackageResponse_packageName :: Lens' DescribePackageResponse Text

The package's name.

describePackageResponse_arn :: Lens' DescribePackageResponse Text

The package's ARN.

describePackageResponse_storageLocation :: Lens' DescribePackageResponse StorageLocation

The package's storage location.

describePackageResponse_createdTime :: Lens' DescribePackageResponse UTCTime

When the package was created.

describePackageResponse_tags :: Lens' DescribePackageResponse (HashMap Text Text)

The package's tags.