Amazonka.Panorama.DescribePackageVersion

Description

Returns information about a package version.

Creating a Request

data DescribePackageVersion

See: newDescribePackageVersion smart constructor.

Constructors

DescribePackageVersion'

Fields

patchVersion :: Maybe Text

The version's patch version.

ownerAccount :: Maybe Text

The version's owner account.

packageId :: Text

The version's ID.

packageVersion :: Text

The version's version.

newDescribePackageVersion

Arguments

:: Text
$sel:packageId:DescribePackageVersion'

-> Text
$sel:packageVersion:DescribePackageVersion'

-> DescribePackageVersion

Create a value of DescribePackageVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:patchVersion:DescribePackageVersion', describePackageVersion_patchVersion - The version's patch version.

$sel:ownerAccount:DescribePackageVersion', describePackageVersion_ownerAccount - The version's owner account.

$sel:packageId:DescribePackageVersion', describePackageVersion_packageId - The version's ID.

$sel:packageVersion:DescribePackageVersion', describePackageVersion_packageVersion - The version's version.

Request Lenses

describePackageVersion_patchVersion :: Lens' DescribePackageVersion (Maybe Text)

The version's patch version.

describePackageVersion_ownerAccount :: Lens' DescribePackageVersion (Maybe Text)

The version's owner account.

describePackageVersion_packageId :: Lens' DescribePackageVersion Text

The version's ID.

describePackageVersion_packageVersion :: Lens' DescribePackageVersion Text

The version's version.

Destructuring the Response

data DescribePackageVersionResponse

See: newDescribePackageVersionResponse smart constructor.

Constructors

DescribePackageVersionResponse'

Fields

statusDescription :: Maybe Text

The version's status description.

packageArn :: Maybe Text

The ARN of the package.

registeredTime :: Maybe POSIX

The version's registered time.

ownerAccount :: Maybe Text

The account ID of the version's owner.

httpStatus :: Int

The response's http status code.

packageId :: Text

The version's ID.

packageName :: Text

The version's name.

packageVersion :: Text

The version's version.

patchVersion :: Text

The version's patch version.

isLatestPatch :: Bool

Whether the version is the latest available.

status :: PackageVersionStatus

The version's status.

newDescribePackageVersionResponse

Arguments

:: Int
$sel:httpStatus:DescribePackageVersionResponse'

-> Text
$sel:packageId:DescribePackageVersion'

-> Text
$sel:packageName:DescribePackageVersionResponse'

-> Text
$sel:packageVersion:DescribePackageVersion'

-> Text
$sel:patchVersion:DescribePackageVersion'

-> Bool
$sel:isLatestPatch:DescribePackageVersionResponse'

-> PackageVersionStatus
$sel:status:DescribePackageVersionResponse'

-> DescribePackageVersionResponse

Create a value of DescribePackageVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusDescription:DescribePackageVersionResponse', describePackageVersionResponse_statusDescription - The version's status description.

$sel:packageArn:DescribePackageVersionResponse', describePackageVersionResponse_packageArn - The ARN of the package.

$sel:registeredTime:DescribePackageVersionResponse', describePackageVersionResponse_registeredTime - The version's registered time.

$sel:ownerAccount:DescribePackageVersion', describePackageVersionResponse_ownerAccount - The account ID of the version's owner.

$sel:httpStatus:DescribePackageVersionResponse', describePackageVersionResponse_httpStatus - The response's http status code.

$sel:packageId:DescribePackageVersion', describePackageVersionResponse_packageId - The version's ID.

$sel:packageName:DescribePackageVersionResponse', describePackageVersionResponse_packageName - The version's name.

$sel:packageVersion:DescribePackageVersion', describePackageVersionResponse_packageVersion - The version's version.

$sel:patchVersion:DescribePackageVersion', describePackageVersionResponse_patchVersion - The version's patch version.

$sel:isLatestPatch:DescribePackageVersionResponse', describePackageVersionResponse_isLatestPatch - Whether the version is the latest available.

$sel:status:DescribePackageVersionResponse', describePackageVersionResponse_status - The version's status.

Response Lenses

describePackageVersionResponse_statusDescription :: Lens' DescribePackageVersionResponse (Maybe Text)

The version's status description.

describePackageVersionResponse_packageArn :: Lens' DescribePackageVersionResponse (Maybe Text)

The ARN of the package.

describePackageVersionResponse_registeredTime :: Lens' DescribePackageVersionResponse (Maybe UTCTime)

The version's registered time.

describePackageVersionResponse_ownerAccount :: Lens' DescribePackageVersionResponse (Maybe Text)

The account ID of the version's owner.

describePackageVersionResponse_httpStatus :: Lens' DescribePackageVersionResponse Int

The response's http status code.

describePackageVersionResponse_packageId :: Lens' DescribePackageVersionResponse Text

The version's ID.

describePackageVersionResponse_packageName :: Lens' DescribePackageVersionResponse Text

The version's name.

describePackageVersionResponse_packageVersion :: Lens' DescribePackageVersionResponse Text

The version's version.

describePackageVersionResponse_patchVersion :: Lens' DescribePackageVersionResponse Text

The version's patch version.

describePackageVersionResponse_isLatestPatch :: Lens' DescribePackageVersionResponse Bool

Whether the version is the latest available.

describePackageVersionResponse_status :: Lens' DescribePackageVersionResponse PackageVersionStatus

The version's status.