Amazonka.Panorama.Lens

Description

Operations

UpdateDeviceMetadata

updateDeviceMetadata_description :: Lens' UpdateDeviceMetadata (Maybe Text)

A description for the device.

updateDeviceMetadata_deviceId :: Lens' UpdateDeviceMetadata Text

The device's ID.

updateDeviceMetadataResponse_deviceId :: Lens' UpdateDeviceMetadataResponse (Maybe Text)

The device's ID.

updateDeviceMetadataResponse_httpStatus :: Lens' UpdateDeviceMetadataResponse Int

The response's http status code.

DescribeApplicationInstanceDetails

describeApplicationInstanceDetails_applicationInstanceId :: Lens' DescribeApplicationInstanceDetails Text

The application instance's ID.

describeApplicationInstanceDetailsResponse_createdTime :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe UTCTime)

When the application instance was created.

describeApplicationInstanceDetailsResponse_defaultRuntimeContextDevice :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe Text)

The application instance's default runtime context device.

describeApplicationInstanceDetailsResponse_manifestOverridesPayload :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe ManifestOverridesPayload)

Parameter overrides for the configuration manifest.

describeApplicationInstanceDetailsResponse_name :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe Text)

The application instance's name.

describeApplicationInstanceDetailsResponse_applicationInstanceId :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe Text)

The application instance's ID.

describeApplicationInstanceDetailsResponse_description :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe Text)

The application instance's description.

describeApplicationInstanceDetailsResponse_manifestPayload :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe ManifestPayload)

The application instance's configuration manifest.

describeApplicationInstanceDetailsResponse_applicationInstanceIdToReplace :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe Text)

The ID of the application instance that this instance replaced.

describeApplicationInstanceDetailsResponse_httpStatus :: Lens' DescribeApplicationInstanceDetailsResponse Int

The response's http status code.

CreateApplicationInstance

createApplicationInstance_manifestOverridesPayload :: Lens' CreateApplicationInstance (Maybe ManifestOverridesPayload)

Setting overrides for the application manifest.

createApplicationInstance_name :: Lens' CreateApplicationInstance (Maybe Text)

A name for the application instance.

createApplicationInstance_runtimeRoleArn :: Lens' CreateApplicationInstance (Maybe Text)

The ARN of a runtime role for the application instance.

createApplicationInstance_description :: Lens' CreateApplicationInstance (Maybe Text)

A description for the application instance.

createApplicationInstance_tags :: Lens' CreateApplicationInstance (Maybe (HashMap Text Text))

Tags for the application instance.

createApplicationInstance_applicationInstanceIdToReplace :: Lens' CreateApplicationInstance (Maybe Text)

The ID of an application instance to replace with the new instance.

createApplicationInstance_manifestPayload :: Lens' CreateApplicationInstance ManifestPayload

The application's manifest document.

createApplicationInstance_defaultRuntimeContextDevice :: Lens' CreateApplicationInstance Text

A device's ID.

createApplicationInstanceResponse_httpStatus :: Lens' CreateApplicationInstanceResponse Int

The response's http status code.

createApplicationInstanceResponse_applicationInstanceId :: Lens' CreateApplicationInstanceResponse Text

The application instance's ID.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The resource's ARN.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

A list of tags.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

RemoveApplicationInstance

removeApplicationInstance_applicationInstanceId :: Lens' RemoveApplicationInstance Text

An application instance ID.

removeApplicationInstanceResponse_httpStatus :: Lens' RemoveApplicationInstanceResponse Int

The response's http status code.

ListDevicesJobs

listDevicesJobs_nextToken :: Lens' ListDevicesJobs (Maybe Text)

Specify the pagination token from a previous request to retrieve the next page of results.

listDevicesJobs_deviceId :: Lens' ListDevicesJobs (Maybe Text)

Filter results by the job's target device ID.

listDevicesJobs_maxResults :: Lens' ListDevicesJobs (Maybe Natural)

The maximum number of device jobs to return in one page of results.

listDevicesJobsResponse_nextToken :: Lens' ListDevicesJobsResponse (Maybe Text)

A pagination token that's included if more results are available.

listDevicesJobsResponse_deviceJobs :: Lens' ListDevicesJobsResponse (Maybe [DeviceJob])

A list of jobs.

listDevicesJobsResponse_httpStatus :: Lens' ListDevicesJobsResponse Int

The response's http status code.

CreateNodeFromTemplateJob

createNodeFromTemplateJob_jobTags :: Lens' CreateNodeFromTemplateJob (Maybe [JobResourceTags])

Tags for the job.

createNodeFromTemplateJob_nodeDescription :: Lens' CreateNodeFromTemplateJob (Maybe Text)

A description for the node.

createNodeFromTemplateJob_templateType :: Lens' CreateNodeFromTemplateJob TemplateType

The type of node.

createNodeFromTemplateJob_outputPackageName :: Lens' CreateNodeFromTemplateJob Text

An output package name for the node.

createNodeFromTemplateJob_outputPackageVersion :: Lens' CreateNodeFromTemplateJob Text

An output package version for the node.

createNodeFromTemplateJob_nodeName :: Lens' CreateNodeFromTemplateJob Text

A name for the node.

createNodeFromTemplateJob_templateParameters :: Lens' CreateNodeFromTemplateJob (HashMap Text Text)

Template parameters for the node.

createNodeFromTemplateJobResponse_httpStatus :: Lens' CreateNodeFromTemplateJobResponse Int

The response's http status code.

createNodeFromTemplateJobResponse_jobId :: Lens' CreateNodeFromTemplateJobResponse Text

The job's ID.

DeregisterPackageVersion

deregisterPackageVersion_updatedLatestPatchVersion :: Lens' DeregisterPackageVersion (Maybe Text)

If the version was marked latest, the new version to maker as latest.

deregisterPackageVersion_ownerAccount :: Lens' DeregisterPackageVersion (Maybe Text)

An owner account.

deregisterPackageVersion_packageId :: Lens' DeregisterPackageVersion Text

A package ID.

deregisterPackageVersion_packageVersion :: Lens' DeregisterPackageVersion Text

A package version.

deregisterPackageVersion_patchVersion :: Lens' DeregisterPackageVersion Text

A patch version.

deregisterPackageVersionResponse_httpStatus :: Lens' DeregisterPackageVersionResponse Int

The response's http status code.

ListPackages

listPackages_nextToken :: Lens' ListPackages (Maybe Text)

Specify the pagination token from a previous request to retrieve the next page of results.

listPackages_maxResults :: Lens' ListPackages (Maybe Natural)

The maximum number of packages to return in one page of results.

listPackagesResponse_packages :: Lens' ListPackagesResponse (Maybe [PackageListItem])

A list of packages.

listPackagesResponse_nextToken :: Lens' ListPackagesResponse (Maybe Text)

A pagination token that's included if more results are available.

listPackagesResponse_httpStatus :: Lens' ListPackagesResponse Int

The response's http status code.

DescribeApplicationInstance

describeApplicationInstance_applicationInstanceId :: Lens' DescribeApplicationInstance Text

The application instance's ID.

describeApplicationInstanceResponse_status :: Lens' DescribeApplicationInstanceResponse (Maybe ApplicationInstanceStatus)

The application instance's status.

describeApplicationInstanceResponse_statusDescription :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The application instance's status description.

describeApplicationInstanceResponse_lastUpdatedTime :: Lens' DescribeApplicationInstanceResponse (Maybe UTCTime)

The application instance was updated.

describeApplicationInstanceResponse_arn :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The application instance's ARN.

describeApplicationInstanceResponse_createdTime :: Lens' DescribeApplicationInstanceResponse (Maybe UTCTime)

When the application instance was created.

describeApplicationInstanceResponse_defaultRuntimeContextDevice :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The device's ID.

describeApplicationInstanceResponse_defaultRuntimeContextDeviceName :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The device's bane.

describeApplicationInstanceResponse_name :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The application instance's name.

describeApplicationInstanceResponse_runtimeRoleArn :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The application instance's runtime role ARN.

describeApplicationInstanceResponse_healthStatus :: Lens' DescribeApplicationInstanceResponse (Maybe ApplicationInstanceHealthStatus)

The application instance's health status.

describeApplicationInstanceResponse_applicationInstanceId :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The application instance's ID.

describeApplicationInstanceResponse_description :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The application instance's description.

describeApplicationInstanceResponse_tags :: Lens' DescribeApplicationInstanceResponse (Maybe (HashMap Text Text))

The application instance's tags.

describeApplicationInstanceResponse_applicationInstanceIdToReplace :: Lens' DescribeApplicationInstanceResponse (Maybe Text)

The ID of the application instance that this instance replaced.

describeApplicationInstanceResponse_httpStatus :: Lens' DescribeApplicationInstanceResponse Int

The response's http status code.

RegisterPackageVersion

registerPackageVersion_markLatest :: Lens' RegisterPackageVersion (Maybe Bool)

Whether to mark the new version as the latest version.

registerPackageVersion_ownerAccount :: Lens' RegisterPackageVersion (Maybe Text)

An owner account.

registerPackageVersion_packageId :: Lens' RegisterPackageVersion Text

A package ID.

registerPackageVersion_packageVersion :: Lens' RegisterPackageVersion Text

A package version.

registerPackageVersion_patchVersion :: Lens' RegisterPackageVersion Text

A patch version.

registerPackageVersionResponse_httpStatus :: Lens' RegisterPackageVersionResponse Int

The response's http status code.

DescribeNodeFromTemplateJob

describeNodeFromTemplateJob_jobId :: Lens' DescribeNodeFromTemplateJob Text

The job's ID.

describeNodeFromTemplateJobResponse_jobTags :: Lens' DescribeNodeFromTemplateJobResponse (Maybe [JobResourceTags])

The job's tags.

describeNodeFromTemplateJobResponse_nodeDescription :: Lens' DescribeNodeFromTemplateJobResponse (Maybe Text)

The node's description.

describeNodeFromTemplateJobResponse_httpStatus :: Lens' DescribeNodeFromTemplateJobResponse Int

The response's http status code.

describeNodeFromTemplateJobResponse_jobId :: Lens' DescribeNodeFromTemplateJobResponse Text

The job's ID.

describeNodeFromTemplateJobResponse_status :: Lens' DescribeNodeFromTemplateJobResponse NodeFromTemplateJobStatus

The job's status.

describeNodeFromTemplateJobResponse_statusMessage :: Lens' DescribeNodeFromTemplateJobResponse Text

The job's status message.

describeNodeFromTemplateJobResponse_createdTime :: Lens' DescribeNodeFromTemplateJobResponse UTCTime

When the job was created.

describeNodeFromTemplateJobResponse_lastUpdatedTime :: Lens' DescribeNodeFromTemplateJobResponse UTCTime

When the job was updated.

describeNodeFromTemplateJobResponse_outputPackageName :: Lens' DescribeNodeFromTemplateJobResponse Text

The job's output package name.

describeNodeFromTemplateJobResponse_outputPackageVersion :: Lens' DescribeNodeFromTemplateJobResponse Text

The job's output package version.

describeNodeFromTemplateJobResponse_nodeName :: Lens' DescribeNodeFromTemplateJobResponse Text

The node's name.

describeNodeFromTemplateJobResponse_templateType :: Lens' DescribeNodeFromTemplateJobResponse TemplateType

The job's template type.

describeNodeFromTemplateJobResponse_templateParameters :: Lens' DescribeNodeFromTemplateJobResponse (HashMap Text Text)

The job's template parameters.

CreatePackageImportJob

createPackageImportJob_jobTags :: Lens' CreatePackageImportJob (Maybe [JobResourceTags])

Tags for the package import job.

createPackageImportJob_jobType :: Lens' CreatePackageImportJob PackageImportJobType

A job type for the package import job.

createPackageImportJob_inputConfig :: Lens' CreatePackageImportJob PackageImportJobInputConfig

An input config for the package import job.

createPackageImportJob_outputConfig :: Lens' CreatePackageImportJob PackageImportJobOutputConfig

An output config for the package import job.

createPackageImportJob_clientToken :: Lens' CreatePackageImportJob Text

A client token for the package import job.

createPackageImportJobResponse_httpStatus :: Lens' CreatePackageImportJobResponse Int

The response's http status code.

createPackageImportJobResponse_jobId :: Lens' CreatePackageImportJobResponse Text

The job's ID.

DescribePackage

describePackage_packageId :: Lens' DescribePackage Text

The package's ID.

describePackageResponse_writeAccessPrincipalArns :: Lens' DescribePackageResponse (Maybe [Text])

ARNs of accounts that have write access to the package.

describePackageResponse_readAccessPrincipalArns :: Lens' DescribePackageResponse (Maybe [Text])

ARNs of accounts that have read access to the package.

describePackageResponse_httpStatus :: Lens' DescribePackageResponse Int

The response's http status code.

describePackageResponse_packageId :: Lens' DescribePackageResponse Text

The package's ID.

describePackageResponse_packageName :: Lens' DescribePackageResponse Text

The package's name.

describePackageResponse_arn :: Lens' DescribePackageResponse Text

The package's ARN.

describePackageResponse_storageLocation :: Lens' DescribePackageResponse StorageLocation

The package's storage location.

describePackageResponse_createdTime :: Lens' DescribePackageResponse UTCTime

When the package was created.

describePackageResponse_tags :: Lens' DescribePackageResponse (HashMap Text Text)

The package's tags.

ListApplicationInstances

listApplicationInstances_nextToken :: Lens' ListApplicationInstances (Maybe Text)

Specify the pagination token from a previous request to retrieve the next page of results.

listApplicationInstances_statusFilter :: Lens' ListApplicationInstances (Maybe StatusFilter)

Only include instances with a specific status.

listApplicationInstances_deviceId :: Lens' ListApplicationInstances (Maybe Text)

The application instances' device ID.

listApplicationInstances_maxResults :: Lens' ListApplicationInstances (Maybe Natural)

The maximum number of application instances to return in one page of results.

listApplicationInstancesResponse_nextToken :: Lens' ListApplicationInstancesResponse (Maybe Text)

A pagination token that's included if more results are available.

listApplicationInstancesResponse_applicationInstances :: Lens' ListApplicationInstancesResponse (Maybe [ApplicationInstance])

A list of application instances.

listApplicationInstancesResponse_httpStatus :: Lens' ListApplicationInstancesResponse Int

The response's http status code.

DescribeDeviceJob

describeDeviceJob_jobId :: Lens' DescribeDeviceJob Text

The job's ID.

describeDeviceJobResponse_status :: Lens' DescribeDeviceJobResponse (Maybe UpdateProgress)

The job's status.

describeDeviceJobResponse_jobId :: Lens' DescribeDeviceJobResponse (Maybe Text)

The job's ID.

describeDeviceJobResponse_createdTime :: Lens' DescribeDeviceJobResponse (Maybe UTCTime)

When the job was created.

describeDeviceJobResponse_deviceArn :: Lens' DescribeDeviceJobResponse (Maybe Text)

The device's ARN.

describeDeviceJobResponse_imageVersion :: Lens' DescribeDeviceJobResponse (Maybe Text)

For an OTA job, the target version of the device software.

describeDeviceJobResponse_deviceName :: Lens' DescribeDeviceJobResponse (Maybe Text)

The device's name.

describeDeviceJobResponse_deviceId :: Lens' DescribeDeviceJobResponse (Maybe Text)

The device's ID.

describeDeviceJobResponse_deviceType :: Lens' DescribeDeviceJobResponse (Maybe DeviceType)

The device's type.

describeDeviceJobResponse_httpStatus :: Lens' DescribeDeviceJobResponse Int

The response's http status code.

DescribePackageImportJob

describePackageImportJob_jobId :: Lens' DescribePackageImportJob Text

The job's ID.

describePackageImportJobResponse_clientToken :: Lens' DescribePackageImportJobResponse (Maybe Text)

The job's client token.

describePackageImportJobResponse_jobTags :: Lens' DescribePackageImportJobResponse (Maybe [JobResourceTags])

The job's tags.

describePackageImportJobResponse_httpStatus :: Lens' DescribePackageImportJobResponse Int

The response's http status code.

describePackageImportJobResponse_jobId :: Lens' DescribePackageImportJobResponse Text

The job's ID.

describePackageImportJobResponse_jobType :: Lens' DescribePackageImportJobResponse PackageImportJobType

The job's type.

describePackageImportJobResponse_inputConfig :: Lens' DescribePackageImportJobResponse PackageImportJobInputConfig

The job's input config.

describePackageImportJobResponse_outputConfig :: Lens' DescribePackageImportJobResponse PackageImportJobOutputConfig

The job's output config.

describePackageImportJobResponse_output :: Lens' DescribePackageImportJobResponse PackageImportJobOutput

The job's output.

describePackageImportJobResponse_createdTime :: Lens' DescribePackageImportJobResponse UTCTime

When the job was created.

describePackageImportJobResponse_lastUpdatedTime :: Lens' DescribePackageImportJobResponse UTCTime

When the job was updated.

describePackageImportJobResponse_status :: Lens' DescribePackageImportJobResponse PackageImportJobStatus

The job's status.

describePackageImportJobResponse_statusMessage :: Lens' DescribePackageImportJobResponse Text

The job's status message.

DescribeDevice

describeDevice_deviceId :: Lens' DescribeDevice Text

The device's ID.

describeDeviceResponse_latestSoftware :: Lens' DescribeDeviceResponse (Maybe Text)

The latest software version available for the device.

describeDeviceResponse_provisioningStatus :: Lens' DescribeDeviceResponse (Maybe DeviceStatus)

The device's provisioning status.

describeDeviceResponse_arn :: Lens' DescribeDeviceResponse (Maybe Text)

The device's ARN.

describeDeviceResponse_createdTime :: Lens' DescribeDeviceResponse (Maybe UTCTime)

When the device was created.

describeDeviceResponse_currentSoftware :: Lens' DescribeDeviceResponse (Maybe Text)

The device's current software version.

describeDeviceResponse_name :: Lens' DescribeDeviceResponse (Maybe Text)

The device's name.

describeDeviceResponse_deviceConnectionStatus :: Lens' DescribeDeviceResponse (Maybe DeviceConnectionStatus)

The device's connection status.

describeDeviceResponse_deviceId :: Lens' DescribeDeviceResponse (Maybe Text)

The device's ID.

describeDeviceResponse_type :: Lens' DescribeDeviceResponse (Maybe DeviceType)

The device's type.

describeDeviceResponse_leaseExpirationTime :: Lens' DescribeDeviceResponse (Maybe UTCTime)

The device's lease expiration time.

describeDeviceResponse_serialNumber :: Lens' DescribeDeviceResponse (Maybe Text)

The device's serial number.

describeDeviceResponse_currentNetworkingStatus :: Lens' DescribeDeviceResponse (Maybe NetworkStatus)

The device's networking status.

describeDeviceResponse_description :: Lens' DescribeDeviceResponse (Maybe Text)

The device's description.

describeDeviceResponse_networkingConfiguration :: Lens' DescribeDeviceResponse (Maybe NetworkPayload)

The device's networking configuration.

describeDeviceResponse_tags :: Lens' DescribeDeviceResponse (Maybe (HashMap Text Text))

The device's tags.

describeDeviceResponse_httpStatus :: Lens' DescribeDeviceResponse Int

The response's http status code.

DescribePackageVersion

describePackageVersion_patchVersion :: Lens' DescribePackageVersion (Maybe Text)

The version's patch version.

describePackageVersion_ownerAccount :: Lens' DescribePackageVersion (Maybe Text)

The version's owner account.

describePackageVersion_packageId :: Lens' DescribePackageVersion Text

The version's ID.

describePackageVersion_packageVersion :: Lens' DescribePackageVersion Text

The version's version.

describePackageVersionResponse_statusDescription :: Lens' DescribePackageVersionResponse (Maybe Text)

The version's status description.

describePackageVersionResponse_packageArn :: Lens' DescribePackageVersionResponse (Maybe Text)

The ARN of the package.

describePackageVersionResponse_registeredTime :: Lens' DescribePackageVersionResponse (Maybe UTCTime)

The version's registered time.

describePackageVersionResponse_ownerAccount :: Lens' DescribePackageVersionResponse (Maybe Text)

The account ID of the version's owner.

describePackageVersionResponse_httpStatus :: Lens' DescribePackageVersionResponse Int

The response's http status code.

describePackageVersionResponse_packageId :: Lens' DescribePackageVersionResponse Text

The version's ID.

describePackageVersionResponse_packageName :: Lens' DescribePackageVersionResponse Text

The version's name.

describePackageVersionResponse_packageVersion :: Lens' DescribePackageVersionResponse Text

The version's version.

describePackageVersionResponse_patchVersion :: Lens' DescribePackageVersionResponse Text

The version's patch version.

describePackageVersionResponse_isLatestPatch :: Lens' DescribePackageVersionResponse Bool

Whether the version is the latest available.

describePackageVersionResponse_status :: Lens' DescribePackageVersionResponse PackageVersionStatus

The version's status.

DescribeNode

describeNode_ownerAccount :: Lens' DescribeNode (Maybe Text)

The account ID of the node's owner.

describeNode_nodeId :: Lens' DescribeNode Text

The node's ID.

describeNodeResponse_assetName :: Lens' DescribeNodeResponse (Maybe Text)

The node's asset name.

describeNodeResponse_packageArn :: Lens' DescribeNodeResponse (Maybe Text)

The node's ARN.

describeNodeResponse_httpStatus :: Lens' DescribeNodeResponse Int

The response's http status code.

describeNodeResponse_nodeId :: Lens' DescribeNodeResponse Text

The node's ID.

describeNodeResponse_name :: Lens' DescribeNodeResponse Text

The node's name.

describeNodeResponse_category :: Lens' DescribeNodeResponse NodeCategory

The node's category.

describeNodeResponse_ownerAccount :: Lens' DescribeNodeResponse Text

The account ID of the node's owner.

describeNodeResponse_packageName :: Lens' DescribeNodeResponse Text

The node's package name.

describeNodeResponse_packageId :: Lens' DescribeNodeResponse Text

The node's package ID.

describeNodeResponse_packageVersion :: Lens' DescribeNodeResponse Text

The node's package version.

describeNodeResponse_patchVersion :: Lens' DescribeNodeResponse Text

The node's patch version.

describeNodeResponse_nodeInterface :: Lens' DescribeNodeResponse NodeInterface

The node's interface.

describeNodeResponse_description :: Lens' DescribeNodeResponse Text

The node's description.

describeNodeResponse_createdTime :: Lens' DescribeNodeResponse UTCTime

When the node was created.

describeNodeResponse_lastUpdatedTime :: Lens' DescribeNodeResponse UTCTime

When the node was updated.

ListNodeFromTemplateJobs

listNodeFromTemplateJobs_nextToken :: Lens' ListNodeFromTemplateJobs (Maybe Text)

Specify the pagination token from a previous request to retrieve the next page of results.

listNodeFromTemplateJobs_maxResults :: Lens' ListNodeFromTemplateJobs (Maybe Natural)

The maximum number of node from template jobs to return in one page of results.

listNodeFromTemplateJobsResponse_nextToken :: Lens' ListNodeFromTemplateJobsResponse (Maybe Text)

A pagination token that's included if more results are available.

listNodeFromTemplateJobsResponse_httpStatus :: Lens' ListNodeFromTemplateJobsResponse Int

The response's http status code.

listNodeFromTemplateJobsResponse_nodeFromTemplateJobs :: Lens' ListNodeFromTemplateJobsResponse [NodeFromTemplateJob]

A list of jobs.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The resource's ARN.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

Tags for the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListApplicationInstanceDependencies

listApplicationInstanceDependencies_nextToken :: Lens' ListApplicationInstanceDependencies (Maybe Text)

Specify the pagination token from a previous request to retrieve the next page of results.

listApplicationInstanceDependencies_maxResults :: Lens' ListApplicationInstanceDependencies (Maybe Natural)

The maximum number of application instance dependencies to return in one page of results.

listApplicationInstanceDependencies_applicationInstanceId :: Lens' ListApplicationInstanceDependencies Text

The application instance's ID.

listApplicationInstanceDependenciesResponse_packageObjects :: Lens' ListApplicationInstanceDependenciesResponse (Maybe [PackageObject])

A list of package objects.

listApplicationInstanceDependenciesResponse_nextToken :: Lens' ListApplicationInstanceDependenciesResponse (Maybe Text)

A pagination token that's included if more results are available.

listApplicationInstanceDependenciesResponse_httpStatus :: Lens' ListApplicationInstanceDependenciesResponse Int

The response's http status code.

ListApplicationInstanceNodeInstances

listApplicationInstanceNodeInstances_nextToken :: Lens' ListApplicationInstanceNodeInstances (Maybe Text)

Specify the pagination token from a previous request to retrieve the next page of results.

listApplicationInstanceNodeInstances_maxResults :: Lens' ListApplicationInstanceNodeInstances (Maybe Natural)

The maximum number of node instances to return in one page of results.

listApplicationInstanceNodeInstances_applicationInstanceId :: Lens' ListApplicationInstanceNodeInstances Text

The node instances' application instance ID.

listApplicationInstanceNodeInstancesResponse_nextToken :: Lens' ListApplicationInstanceNodeInstancesResponse (Maybe Text)

A pagination token that's included if more results are available.

listApplicationInstanceNodeInstancesResponse_nodeInstances :: Lens' ListApplicationInstanceNodeInstancesResponse (Maybe [NodeInstance])

A list of node instances.

listApplicationInstanceNodeInstancesResponse_httpStatus :: Lens' ListApplicationInstanceNodeInstancesResponse Int

The response's http status code.

CreateJobForDevices

createJobForDevices_deviceIds :: Lens' CreateJobForDevices (NonEmpty Text)

IDs of target devices.

createJobForDevices_deviceJobConfig :: Lens' CreateJobForDevices DeviceJobConfig

Configuration settings for the job.

createJobForDevices_jobType :: Lens' CreateJobForDevices JobType

The type of job to run.

createJobForDevicesResponse_httpStatus :: Lens' CreateJobForDevicesResponse Int

The response's http status code.

createJobForDevicesResponse_jobs :: Lens' CreateJobForDevicesResponse [Job]

A list of jobs.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The resource's ARN.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

Tag keys to remove.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DeletePackage

deletePackage_forceDelete :: Lens' DeletePackage (Maybe Bool)

Delete the package even if it has artifacts stored in its access point. Deletes the package's artifacts from Amazon S3.

deletePackage_packageId :: Lens' DeletePackage Text

The package's ID.

deletePackageResponse_httpStatus :: Lens' DeletePackageResponse Int

The response's http status code.

CreatePackage

createPackage_tags :: Lens' CreatePackage (Maybe (HashMap Text Text))

Tags for the package.

createPackage_packageName :: Lens' CreatePackage Text

A name for the package.

createPackageResponse_packageId :: Lens' CreatePackageResponse (Maybe Text)

The package's ID.

createPackageResponse_arn :: Lens' CreatePackageResponse (Maybe Text)

The package's ARN.

createPackageResponse_httpStatus :: Lens' CreatePackageResponse Int

The response's http status code.

createPackageResponse_storageLocation :: Lens' CreatePackageResponse StorageLocation

The package's storage location.

ProvisionDevice

provisionDevice_description :: Lens' ProvisionDevice (Maybe Text)

A description for the device.

provisionDevice_networkingConfiguration :: Lens' ProvisionDevice (Maybe NetworkPayload)

A networking configuration for the device.

provisionDevice_tags :: Lens' ProvisionDevice (Maybe (HashMap Text Text))

Tags for the device.

provisionDevice_name :: Lens' ProvisionDevice Text

A name for the device.

provisionDeviceResponse_certificates :: Lens' ProvisionDeviceResponse (Maybe ByteString)

The device's configuration bundle.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

provisionDeviceResponse_deviceId :: Lens' ProvisionDeviceResponse (Maybe Text)

The device's ID.

provisionDeviceResponse_iotThingName :: Lens' ProvisionDeviceResponse (Maybe Text)

The device's IoT thing name.

provisionDeviceResponse_httpStatus :: Lens' ProvisionDeviceResponse Int

The response's http status code.

provisionDeviceResponse_arn :: Lens' ProvisionDeviceResponse Text

The device's ARN.

provisionDeviceResponse_status :: Lens' ProvisionDeviceResponse DeviceStatus

The device's status.

DeleteDevice

deleteDevice_deviceId :: Lens' DeleteDevice Text

The device's ID.

deleteDeviceResponse_deviceId :: Lens' DeleteDeviceResponse (Maybe Text)

The device's ID.

deleteDeviceResponse_httpStatus :: Lens' DeleteDeviceResponse Int

The response's http status code.

ListNodes

listNodes_patchVersion :: Lens' ListNodes (Maybe Text)

Search for nodes by patch version.

listNodes_packageName :: Lens' ListNodes (Maybe Text)

Search for nodes by name.

listNodes_packageVersion :: Lens' ListNodes (Maybe Text)

Search for nodes by version.

listNodes_category :: Lens' ListNodes (Maybe NodeCategory)

Search for nodes by category.

listNodes_nextToken :: Lens' ListNodes (Maybe Text)

Specify the pagination token from a previous request to retrieve the next page of results.

listNodes_ownerAccount :: Lens' ListNodes (Maybe Text)

Search for nodes by the account ID of the nodes' owner.

listNodes_maxResults :: Lens' ListNodes (Maybe Natural)

The maximum number of nodes to return in one page of results.

listNodesResponse_nextToken :: Lens' ListNodesResponse (Maybe Text)

A pagination token that's included if more results are available.

listNodesResponse_nodes :: Lens' ListNodesResponse (Maybe [Node])

A list of nodes.

listNodesResponse_httpStatus :: Lens' ListNodesResponse Int

The response's http status code.

ListDevices

listDevices_nextToken :: Lens' ListDevices (Maybe Text)

Specify the pagination token from a previous request to retrieve the next page of results.

listDevices_maxResults :: Lens' ListDevices (Maybe Natural)

The maximum number of devices to return in one page of results.

listDevicesResponse_nextToken :: Lens' ListDevicesResponse (Maybe Text)

A pagination token that's included if more results are available.

listDevicesResponse_httpStatus :: Lens' ListDevicesResponse Int

The response's http status code.

listDevicesResponse_devices :: Lens' ListDevicesResponse [Device]

A list of devices.

ListPackageImportJobs

listPackageImportJobs_nextToken :: Lens' ListPackageImportJobs (Maybe Text)

Specify the pagination token from a previous request to retrieve the next page of results.

listPackageImportJobs_maxResults :: Lens' ListPackageImportJobs (Maybe Natural)

The maximum number of package import jobs to return in one page of results.

listPackageImportJobsResponse_nextToken :: Lens' ListPackageImportJobsResponse (Maybe Text)

A pagination token that's included if more results are available.

listPackageImportJobsResponse_httpStatus :: Lens' ListPackageImportJobsResponse Int

The response's http status code.

listPackageImportJobsResponse_packageImportJobs :: Lens' ListPackageImportJobsResponse [PackageImportJob]

A list of package import jobs.

Types

ApplicationInstance

applicationInstance_status :: Lens' ApplicationInstance (Maybe ApplicationInstanceStatus)

The application instance's status.

applicationInstance_statusDescription :: Lens' ApplicationInstance (Maybe Text)

The application instance's status description.

applicationInstance_arn :: Lens' ApplicationInstance (Maybe Text)

The application instance's ARN.

applicationInstance_createdTime :: Lens' ApplicationInstance (Maybe UTCTime)

When the application instance was created.

applicationInstance_defaultRuntimeContextDevice :: Lens' ApplicationInstance (Maybe Text)

The device's ID.

applicationInstance_defaultRuntimeContextDeviceName :: Lens' ApplicationInstance (Maybe Text)

The device's name.

applicationInstance_name :: Lens' ApplicationInstance (Maybe Text)

The application instance's name.

applicationInstance_healthStatus :: Lens' ApplicationInstance (Maybe ApplicationInstanceHealthStatus)

The application instance's health status.

applicationInstance_applicationInstanceId :: Lens' ApplicationInstance (Maybe Text)

The application instance's ID.

applicationInstance_description :: Lens' ApplicationInstance (Maybe Text)

The application instance's description.

applicationInstance_tags :: Lens' ApplicationInstance (Maybe (HashMap Text Text))

The application instance's tags.

Device

device_lastUpdatedTime :: Lens' Device (Maybe UTCTime)

When the device was updated.

device_provisioningStatus :: Lens' Device (Maybe DeviceStatus)

The device's provisioning status.

device_createdTime :: Lens' Device (Maybe UTCTime)

When the device was created.

device_name :: Lens' Device (Maybe Text)

The device's name.

device_deviceId :: Lens' Device (Maybe Text)

The device's ID.

device_leaseExpirationTime :: Lens' Device (Maybe UTCTime)

The device's lease expiration time.

DeviceJob

deviceJob_jobId :: Lens' DeviceJob (Maybe Text)

The job's ID.

deviceJob_createdTime :: Lens' DeviceJob (Maybe UTCTime)

When the job was created.

deviceJob_deviceName :: Lens' DeviceJob (Maybe Text)

The name of the target device

deviceJob_deviceId :: Lens' DeviceJob (Maybe Text)

The ID of the target device.

DeviceJobConfig

deviceJobConfig_oTAJobConfig :: Lens' DeviceJobConfig (Maybe OTAJobConfig)

A configuration for an over-the-air (OTA) upgrade. Required for OTA jobs.

EthernetPayload

ethernetPayload_staticIpConnectionInfo :: Lens' EthernetPayload (Maybe StaticIpConnectionInfo)

Network configuration for a static IP connection.

ethernetPayload_connectionType :: Lens' EthernetPayload ConnectionType

How the device gets an IP address.

EthernetStatus

ethernetStatus_ipAddress :: Lens' EthernetStatus (Maybe Text)

The device's IP address.

ethernetStatus_connectionStatus :: Lens' EthernetStatus (Maybe NetworkConnectionStatus)

The device's connection status.

ethernetStatus_hwAddress :: Lens' EthernetStatus (Maybe Text)

The device's physical address.

Job

job_jobId :: Lens' Job (Maybe Text)

The job's ID.

job_deviceId :: Lens' Job (Maybe Text)

The target device's ID.

JobResourceTags

jobResourceTags_resourceType :: Lens' JobResourceTags JobResourceType

The job's type.

jobResourceTags_tags :: Lens' JobResourceTags (HashMap Text Text)

The job's tags.

ManifestOverridesPayload

manifestOverridesPayload_payloadData :: Lens' ManifestOverridesPayload (Maybe Text)

The overrides document.

ManifestPayload

manifestPayload_payloadData :: Lens' ManifestPayload (Maybe Text)

The application manifest.

NetworkPayload

networkPayload_ethernet1 :: Lens' NetworkPayload (Maybe EthernetPayload)

Settings for Ethernet port 1.

networkPayload_ethernet0 :: Lens' NetworkPayload (Maybe EthernetPayload)

Settings for Ethernet port 0.

NetworkStatus

networkStatus_ethernet1Status :: Lens' NetworkStatus (Maybe EthernetStatus)

The status of Ethernet port 1.

networkStatus_ethernet0Status :: Lens' NetworkStatus (Maybe EthernetStatus)

The status of Ethernet port 0.

Node

node_packageArn :: Lens' Node (Maybe Text)

The node's ARN.

node_ownerAccount :: Lens' Node (Maybe Text)

The account ID of the node's owner.

node_description :: Lens' Node (Maybe Text)

The node's description.

node_nodeId :: Lens' Node Text

The node's ID.

node_name :: Lens' Node Text

The node's name.

node_category :: Lens' Node NodeCategory

The node's category.

node_packageName :: Lens' Node Text

The node's package name.

node_packageId :: Lens' Node Text

The node's package ID.

node_packageVersion :: Lens' Node Text

The node's package version.

node_patchVersion :: Lens' Node Text

The node's patch version.

node_createdTime :: Lens' Node UTCTime

When the node was created.

NodeFromTemplateJob

nodeFromTemplateJob_status :: Lens' NodeFromTemplateJob (Maybe NodeFromTemplateJobStatus)

The job's status.

nodeFromTemplateJob_jobId :: Lens' NodeFromTemplateJob (Maybe Text)

The job's ID.

nodeFromTemplateJob_createdTime :: Lens' NodeFromTemplateJob (Maybe UTCTime)

When the job was created.

nodeFromTemplateJob_templateType :: Lens' NodeFromTemplateJob (Maybe TemplateType)

The job's template type.

nodeFromTemplateJob_nodeName :: Lens' NodeFromTemplateJob (Maybe Text)

The node's name.

nodeFromTemplateJob_statusMessage :: Lens' NodeFromTemplateJob (Maybe Text)

The job's status message.

NodeInputPort

nodeInputPort_maxConnections :: Lens' NodeInputPort (Maybe Int)

The input port's max connections.

nodeInputPort_name :: Lens' NodeInputPort (Maybe Text)

The input port's name.

nodeInputPort_defaultValue :: Lens' NodeInputPort (Maybe Text)

The input port's default value.

nodeInputPort_type :: Lens' NodeInputPort (Maybe PortType)

The input port's type.

nodeInputPort_description :: Lens' NodeInputPort (Maybe Text)

The input port's description.

NodeInstance

nodeInstance_packageName :: Lens' NodeInstance (Maybe Text)

The instance's package name.

nodeInstance_packageVersion :: Lens' NodeInstance (Maybe Text)

The instance's package version.

nodeInstance_packagePatchVersion :: Lens' NodeInstance (Maybe Text)

The instance's package patch version.

nodeInstance_nodeName :: Lens' NodeInstance (Maybe Text)

The instance's name.

nodeInstance_nodeId :: Lens' NodeInstance (Maybe Text)

The node's ID.

nodeInstance_nodeInstanceId :: Lens' NodeInstance Text

The instance's ID.

nodeInstance_currentStatus :: Lens' NodeInstance NodeInstanceStatus

The instance's current status.

NodeInterface

nodeInterface_inputs :: Lens' NodeInterface [NodeInputPort]

The node interface's inputs.

nodeInterface_outputs :: Lens' NodeInterface [NodeOutputPort]

The node interface's outputs.

NodeOutputPort

nodeOutputPort_name :: Lens' NodeOutputPort (Maybe Text)

The output port's name.

nodeOutputPort_type :: Lens' NodeOutputPort (Maybe PortType)

The output port's type.

nodeOutputPort_description :: Lens' NodeOutputPort (Maybe Text)

The output port's description.

OTAJobConfig

oTAJobConfig_imageVersion :: Lens' OTAJobConfig Text

The target version of the device software.

OutPutS3Location

outPutS3Location_bucketName :: Lens' OutPutS3Location Text

The object's bucket.

outPutS3Location_objectKey :: Lens' OutPutS3Location Text

The object's key.

PackageImportJob

packageImportJob_status :: Lens' PackageImportJob (Maybe PackageImportJobStatus)

The job's status.

packageImportJob_jobType :: Lens' PackageImportJob (Maybe PackageImportJobType)

The job's type.

packageImportJob_lastUpdatedTime :: Lens' PackageImportJob (Maybe UTCTime)

When the job was updated.

packageImportJob_jobId :: Lens' PackageImportJob (Maybe Text)

The job's ID.

packageImportJob_createdTime :: Lens' PackageImportJob (Maybe UTCTime)

When the job was created.

packageImportJob_statusMessage :: Lens' PackageImportJob (Maybe Text)

The job's status message.

PackageImportJobInputConfig

packageImportJobInputConfig_packageVersionInputConfig :: Lens' PackageImportJobInputConfig (Maybe PackageVersionInputConfig)

The package version's input configuration.

PackageImportJobOutput

packageImportJobOutput_packageId :: Lens' PackageImportJobOutput Text

The package's ID.

packageImportJobOutput_packageVersion :: Lens' PackageImportJobOutput Text

The package's version.

packageImportJobOutput_patchVersion :: Lens' PackageImportJobOutput Text

The package's patch version.

packageImportJobOutput_outputS3Location :: Lens' PackageImportJobOutput OutPutS3Location

The package's output location.

PackageImportJobOutputConfig

packageImportJobOutputConfig_packageVersionOutputConfig :: Lens' PackageImportJobOutputConfig (Maybe PackageVersionOutputConfig)

The package version's output configuration.

PackageListItem

packageListItem_packageId :: Lens' PackageListItem (Maybe Text)

The package's ID.

packageListItem_arn :: Lens' PackageListItem (Maybe Text)

The package's ARN.

packageListItem_createdTime :: Lens' PackageListItem (Maybe UTCTime)

When the package was created.

packageListItem_packageName :: Lens' PackageListItem (Maybe Text)

The package's name.

packageListItem_tags :: Lens' PackageListItem (Maybe (HashMap Text Text))

The package's tags.

PackageObject

packageObject_name :: Lens' PackageObject Text

The object's name.

packageObject_packageVersion :: Lens' PackageObject Text

The object's package version.

packageObject_patchVersion :: Lens' PackageObject Text

The object's patch version.

PackageVersionInputConfig

packageVersionInputConfig_s3Location :: Lens' PackageVersionInputConfig S3Location

A location in Amazon S3.

PackageVersionOutputConfig

packageVersionOutputConfig_markLatest :: Lens' PackageVersionOutputConfig (Maybe Bool)

Indicates that the version is recommended for all users.

packageVersionOutputConfig_packageName :: Lens' PackageVersionOutputConfig Text

The output's package name.

packageVersionOutputConfig_packageVersion :: Lens' PackageVersionOutputConfig Text

The output's package version.

S3Location

s3Location_region :: Lens' S3Location (Maybe Text)

The bucket's Region.

s3Location_bucketName :: Lens' S3Location Text

A bucket name.

s3Location_objectKey :: Lens' S3Location Text

An object key.

StaticIpConnectionInfo

staticIpConnectionInfo_ipAddress :: Lens' StaticIpConnectionInfo Text

The connection's IP address.

staticIpConnectionInfo_mask :: Lens' StaticIpConnectionInfo Text

The connection's DNS mask.

staticIpConnectionInfo_dns :: Lens' StaticIpConnectionInfo [Text]

The connection's DNS address.

staticIpConnectionInfo_defaultGateway :: Lens' StaticIpConnectionInfo Text

The connection's default gateway.

StorageLocation

storageLocation_bucket :: Lens' StorageLocation Text

The location's bucket.

storageLocation_repoPrefixLocation :: Lens' StorageLocation Text

The location's repo prefix.

storageLocation_generatedPrefixLocation :: Lens' StorageLocation Text

The location's generated prefix.

storageLocation_binaryPrefixLocation :: Lens' StorageLocation Text

The location's binary prefix.

storageLocation_manifestPrefixLocation :: Lens' StorageLocation Text

The location's manifest prefix.