Amazonka.Panorama.Types

Description

Service Configuration

defaultService :: Service

API version 2019-07-24 of the Amazon Panorama SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The request contains an invalid parameter value.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

The requestor does not have permission to access the target action or resource.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The target resource is in use.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request would cause a limit to be exceeded.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error occurred.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The target resource was not found.

ApplicationInstanceHealthStatus

newtype ApplicationInstanceHealthStatus

Constructors

ApplicationInstanceHealthStatus'

Fields

fromApplicationInstanceHealthStatus :: Text

Bundled Patterns

pattern ApplicationInstanceHealthStatus_ERROR :: ApplicationInstanceHealthStatus

pattern ApplicationInstanceHealthStatus_NOT_AVAILABLE :: ApplicationInstanceHealthStatus

pattern ApplicationInstanceHealthStatus_RUNNING :: ApplicationInstanceHealthStatus

ApplicationInstanceStatus

newtype ApplicationInstanceStatus

Constructors

ApplicationInstanceStatus'

Fields

fromApplicationInstanceStatus :: Text

Bundled Patterns

pattern ApplicationInstanceStatus_DEPLOYMENT_ERROR :: ApplicationInstanceStatus

pattern ApplicationInstanceStatus_DEPLOYMENT_IN_PROGRESS :: ApplicationInstanceStatus

pattern ApplicationInstanceStatus_DEPLOYMENT_PENDING :: ApplicationInstanceStatus

pattern ApplicationInstanceStatus_DEPLOYMENT_REQUESTED :: ApplicationInstanceStatus

pattern ApplicationInstanceStatus_DEPLOYMENT_SUCCEEDED :: ApplicationInstanceStatus

pattern ApplicationInstanceStatus_REMOVAL_FAILED :: ApplicationInstanceStatus

pattern ApplicationInstanceStatus_REMOVAL_IN_PROGRESS :: ApplicationInstanceStatus

pattern ApplicationInstanceStatus_REMOVAL_PENDING :: ApplicationInstanceStatus

pattern ApplicationInstanceStatus_REMOVAL_REQUESTED :: ApplicationInstanceStatus

pattern ApplicationInstanceStatus_REMOVAL_SUCCEEDED :: ApplicationInstanceStatus

ConnectionType

newtype ConnectionType

Constructors

ConnectionType'

Fields

fromConnectionType :: Text

Bundled Patterns

pattern ConnectionType_DHCP :: ConnectionType

pattern ConnectionType_STATIC_IP :: ConnectionType

DeviceConnectionStatus

newtype DeviceConnectionStatus

Constructors

DeviceConnectionStatus'

Fields

fromDeviceConnectionStatus :: Text

Bundled Patterns

pattern DeviceConnectionStatus_AWAITING_CREDENTIALS :: DeviceConnectionStatus

pattern DeviceConnectionStatus_ERROR :: DeviceConnectionStatus

pattern DeviceConnectionStatus_NOT_AVAILABLE :: DeviceConnectionStatus

pattern DeviceConnectionStatus_OFFLINE :: DeviceConnectionStatus

pattern DeviceConnectionStatus_ONLINE :: DeviceConnectionStatus

DeviceStatus

newtype DeviceStatus

Constructors

DeviceStatus'

Fields

fromDeviceStatus :: Text

Bundled Patterns

pattern DeviceStatus_AWAITING_PROVISIONING :: DeviceStatus

pattern DeviceStatus_DELETING :: DeviceStatus

pattern DeviceStatus_ERROR :: DeviceStatus

pattern DeviceStatus_FAILED :: DeviceStatus

pattern DeviceStatus_PENDING :: DeviceStatus

pattern DeviceStatus_SUCCEEDED :: DeviceStatus

DeviceType

newtype DeviceType

Constructors

DeviceType'

Fields

fromDeviceType :: Text

Bundled Patterns

pattern DeviceType_PANORAMA_APPLIANCE :: DeviceType

pattern DeviceType_PANORAMA_APPLIANCE_DEVELOPER_KIT :: DeviceType

JobResourceType

newtype JobResourceType

Constructors

JobResourceType'

Fields

fromJobResourceType :: Text

Bundled Patterns

pattern JobResourceType_PACKAGE :: JobResourceType

JobType

newtype JobType

Constructors

JobType'

Fields

fromJobType :: Text

Bundled Patterns

pattern JobType_OTA :: JobType

NetworkConnectionStatus

newtype NetworkConnectionStatus

Constructors

NetworkConnectionStatus'

Fields

fromNetworkConnectionStatus :: Text

Bundled Patterns

pattern NetworkConnectionStatus_CONNECTED :: NetworkConnectionStatus

pattern NetworkConnectionStatus_NOT_CONNECTED :: NetworkConnectionStatus

NodeCategory

newtype NodeCategory

Constructors

NodeCategory'

Fields

fromNodeCategory :: Text

Bundled Patterns

pattern NodeCategory_BUSINESS_LOGIC :: NodeCategory

pattern NodeCategory_MEDIA_SINK :: NodeCategory

pattern NodeCategory_MEDIA_SOURCE :: NodeCategory

pattern NodeCategory_ML_MODEL :: NodeCategory

NodeFromTemplateJobStatus

newtype NodeFromTemplateJobStatus

Constructors

NodeFromTemplateJobStatus'

Fields

fromNodeFromTemplateJobStatus :: Text

Bundled Patterns

pattern NodeFromTemplateJobStatus_FAILED :: NodeFromTemplateJobStatus

pattern NodeFromTemplateJobStatus_PENDING :: NodeFromTemplateJobStatus

pattern NodeFromTemplateJobStatus_SUCCEEDED :: NodeFromTemplateJobStatus

NodeInstanceStatus

newtype NodeInstanceStatus

Constructors

NodeInstanceStatus'

Fields

fromNodeInstanceStatus :: Text

Bundled Patterns

pattern NodeInstanceStatus_ERROR :: NodeInstanceStatus

pattern NodeInstanceStatus_NOT_AVAILABLE :: NodeInstanceStatus

pattern NodeInstanceStatus_RUNNING :: NodeInstanceStatus

PackageImportJobStatus

newtype PackageImportJobStatus

Constructors

PackageImportJobStatus'

Fields

fromPackageImportJobStatus :: Text

Bundled Patterns

pattern PackageImportJobStatus_FAILED :: PackageImportJobStatus

pattern PackageImportJobStatus_PENDING :: PackageImportJobStatus

pattern PackageImportJobStatus_SUCCEEDED :: PackageImportJobStatus

PackageImportJobType

newtype PackageImportJobType

Constructors

PackageImportJobType'

Fields

fromPackageImportJobType :: Text

Bundled Patterns

pattern PackageImportJobType_NODE_PACKAGE_VERSION :: PackageImportJobType

PackageVersionStatus

newtype PackageVersionStatus

Constructors

PackageVersionStatus'

Fields

fromPackageVersionStatus :: Text

Bundled Patterns

pattern PackageVersionStatus_DELETING :: PackageVersionStatus

pattern PackageVersionStatus_FAILED :: PackageVersionStatus

pattern PackageVersionStatus_REGISTER_COMPLETED :: PackageVersionStatus

pattern PackageVersionStatus_REGISTER_PENDING :: PackageVersionStatus

PortType

newtype PortType

Constructors

PortType'

Fields

fromPortType :: Text

Bundled Patterns

pattern PortType_BOOLEAN :: PortType

pattern PortType_FLOAT32 :: PortType

pattern PortType_INT32 :: PortType

pattern PortType_MEDIA :: PortType

pattern PortType_STRING :: PortType

StatusFilter

newtype StatusFilter

Constructors

StatusFilter'

Fields

fromStatusFilter :: Text

Bundled Patterns

pattern StatusFilter_DEPLOYMENT_ERROR :: StatusFilter

pattern StatusFilter_DEPLOYMENT_SUCCEEDED :: StatusFilter

pattern StatusFilter_PROCESSING_DEPLOYMENT :: StatusFilter

pattern StatusFilter_PROCESSING_REMOVAL :: StatusFilter

pattern StatusFilter_REMOVAL_FAILED :: StatusFilter

pattern StatusFilter_REMOVAL_SUCCEEDED :: StatusFilter

TemplateType

newtype TemplateType

Constructors

TemplateType'

Fields

fromTemplateType :: Text

Bundled Patterns

pattern TemplateType_RTSP_CAMERA_STREAM :: TemplateType

UpdateProgress

newtype UpdateProgress

Constructors

UpdateProgress'

Fields

fromUpdateProgress :: Text

Bundled Patterns

pattern UpdateProgress_COMPLETED :: UpdateProgress

pattern UpdateProgress_DOWNLOADING :: UpdateProgress

pattern UpdateProgress_FAILED :: UpdateProgress

pattern UpdateProgress_IN_PROGRESS :: UpdateProgress

pattern UpdateProgress_PENDING :: UpdateProgress

pattern UpdateProgress_REBOOTING :: UpdateProgress

pattern UpdateProgress_VERIFYING :: UpdateProgress

ApplicationInstance

data ApplicationInstance

An application instance on a device.

See: newApplicationInstance smart constructor.

Constructors

ApplicationInstance'

Fields

status :: Maybe ApplicationInstanceStatus

The application instance's status.

statusDescription :: Maybe Text

The application instance's status description.

arn :: Maybe Text

The application instance's ARN.

createdTime :: Maybe POSIX

When the application instance was created.

defaultRuntimeContextDevice :: Maybe Text

The device's ID.

defaultRuntimeContextDeviceName :: Maybe Text

The device's name.

name :: Maybe Text

The application instance's name.

healthStatus :: Maybe ApplicationInstanceHealthStatus

The application instance's health status.

applicationInstanceId :: Maybe Text

The application instance's ID.

description :: Maybe Text

The application instance's description.

tags :: Maybe (HashMap Text Text)

The application instance's tags.

newApplicationInstance :: ApplicationInstance

Create a value of ApplicationInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ApplicationInstance', applicationInstance_status - The application instance's status.

$sel:statusDescription:ApplicationInstance', applicationInstance_statusDescription - The application instance's status description.

$sel:arn:ApplicationInstance', applicationInstance_arn - The application instance's ARN.

$sel:createdTime:ApplicationInstance', applicationInstance_createdTime - When the application instance was created.

$sel:defaultRuntimeContextDevice:ApplicationInstance', applicationInstance_defaultRuntimeContextDevice - The device's ID.

$sel:defaultRuntimeContextDeviceName:ApplicationInstance', applicationInstance_defaultRuntimeContextDeviceName - The device's name.

$sel:name:ApplicationInstance', applicationInstance_name - The application instance's name.

$sel:healthStatus:ApplicationInstance', applicationInstance_healthStatus - The application instance's health status.

$sel:applicationInstanceId:ApplicationInstance', applicationInstance_applicationInstanceId - The application instance's ID.

$sel:description:ApplicationInstance', applicationInstance_description - The application instance's description.

$sel:tags:ApplicationInstance', applicationInstance_tags - The application instance's tags.

applicationInstance_status :: Lens' ApplicationInstance (Maybe ApplicationInstanceStatus)

The application instance's status.

applicationInstance_statusDescription :: Lens' ApplicationInstance (Maybe Text)

The application instance's status description.

applicationInstance_arn :: Lens' ApplicationInstance (Maybe Text)

The application instance's ARN.

applicationInstance_createdTime :: Lens' ApplicationInstance (Maybe UTCTime)

When the application instance was created.

applicationInstance_defaultRuntimeContextDevice :: Lens' ApplicationInstance (Maybe Text)

The device's ID.

applicationInstance_defaultRuntimeContextDeviceName :: Lens' ApplicationInstance (Maybe Text)

The device's name.

applicationInstance_name :: Lens' ApplicationInstance (Maybe Text)

The application instance's name.

applicationInstance_healthStatus :: Lens' ApplicationInstance (Maybe ApplicationInstanceHealthStatus)

The application instance's health status.

applicationInstance_applicationInstanceId :: Lens' ApplicationInstance (Maybe Text)

The application instance's ID.

applicationInstance_description :: Lens' ApplicationInstance (Maybe Text)

The application instance's description.

applicationInstance_tags :: Lens' ApplicationInstance (Maybe (HashMap Text Text))

The application instance's tags.

Device

data Device

A device.

See: newDevice smart constructor.

Constructors

Device'

Fields

lastUpdatedTime :: Maybe POSIX

When the device was updated.

provisioningStatus :: Maybe DeviceStatus

The device's provisioning status.

createdTime :: Maybe POSIX

When the device was created.

name :: Maybe Text

The device's name.

deviceId :: Maybe Text

The device's ID.

leaseExpirationTime :: Maybe POSIX

The device's lease expiration time.

newDevice :: Device

Create a value of Device with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:Device', device_lastUpdatedTime - When the device was updated.

$sel:provisioningStatus:Device', device_provisioningStatus - The device's provisioning status.

$sel:createdTime:Device', device_createdTime - When the device was created.

$sel:name:Device', device_name - The device's name.

$sel:deviceId:Device', device_deviceId - The device's ID.

$sel:leaseExpirationTime:Device', device_leaseExpirationTime - The device's lease expiration time.

device_lastUpdatedTime :: Lens' Device (Maybe UTCTime)

When the device was updated.

device_provisioningStatus :: Lens' Device (Maybe DeviceStatus)

The device's provisioning status.

device_createdTime :: Lens' Device (Maybe UTCTime)

When the device was created.

device_name :: Lens' Device (Maybe Text)

The device's name.

device_deviceId :: Lens' Device (Maybe Text)

The device's ID.

device_leaseExpirationTime :: Lens' Device (Maybe UTCTime)

The device's lease expiration time.

DeviceJob

data DeviceJob

A job that runs on a device.

See: newDeviceJob smart constructor.

Constructors

DeviceJob'

Fields

jobId :: Maybe Text

The job's ID.

createdTime :: Maybe POSIX

When the job was created.

deviceName :: Maybe Text

The name of the target device

deviceId :: Maybe Text

The ID of the target device.

newDeviceJob :: DeviceJob

Create a value of DeviceJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:DeviceJob', deviceJob_jobId - The job's ID.

$sel:createdTime:DeviceJob', deviceJob_createdTime - When the job was created.

$sel:deviceName:DeviceJob', deviceJob_deviceName - The name of the target device

$sel:deviceId:DeviceJob', deviceJob_deviceId - The ID of the target device.

deviceJob_jobId :: Lens' DeviceJob (Maybe Text)

The job's ID.

deviceJob_createdTime :: Lens' DeviceJob (Maybe UTCTime)

When the job was created.

deviceJob_deviceName :: Lens' DeviceJob (Maybe Text)

The name of the target device

deviceJob_deviceId :: Lens' DeviceJob (Maybe Text)

The ID of the target device.

DeviceJobConfig

data DeviceJobConfig

A job's configuration.

See: newDeviceJobConfig smart constructor.

Constructors

DeviceJobConfig'

Fields

oTAJobConfig :: Maybe OTAJobConfig

A configuration for an over-the-air (OTA) upgrade. Required for OTA jobs.

newDeviceJobConfig :: DeviceJobConfig

Create a value of DeviceJobConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:oTAJobConfig:DeviceJobConfig', deviceJobConfig_oTAJobConfig - A configuration for an over-the-air (OTA) upgrade. Required for OTA jobs.

deviceJobConfig_oTAJobConfig :: Lens' DeviceJobConfig (Maybe OTAJobConfig)

A configuration for an over-the-air (OTA) upgrade. Required for OTA jobs.

EthernetPayload

data EthernetPayload

A device's network configuration.

See: newEthernetPayload smart constructor.

Constructors

EthernetPayload'

Fields

staticIpConnectionInfo :: Maybe StaticIpConnectionInfo

Network configuration for a static IP connection.

connectionType :: ConnectionType

How the device gets an IP address.

newEthernetPayload

Arguments

:: ConnectionType
$sel:connectionType:EthernetPayload'

-> EthernetPayload

Create a value of EthernetPayload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:staticIpConnectionInfo:EthernetPayload', ethernetPayload_staticIpConnectionInfo - Network configuration for a static IP connection.

$sel:connectionType:EthernetPayload', ethernetPayload_connectionType - How the device gets an IP address.

ethernetPayload_staticIpConnectionInfo :: Lens' EthernetPayload (Maybe StaticIpConnectionInfo)

Network configuration for a static IP connection.

ethernetPayload_connectionType :: Lens' EthernetPayload ConnectionType

How the device gets an IP address.

EthernetStatus

data EthernetStatus

A device's Ethernet status.

See: newEthernetStatus smart constructor.

Constructors

EthernetStatus'

Fields

ipAddress :: Maybe Text

The device's IP address.

connectionStatus :: Maybe NetworkConnectionStatus

The device's connection status.

hwAddress :: Maybe Text

The device's physical address.

newEthernetStatus :: EthernetStatus

Create a value of EthernetStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddress:EthernetStatus', ethernetStatus_ipAddress - The device's IP address.

$sel:connectionStatus:EthernetStatus', ethernetStatus_connectionStatus - The device's connection status.

$sel:hwAddress:EthernetStatus', ethernetStatus_hwAddress - The device's physical address.

ethernetStatus_ipAddress :: Lens' EthernetStatus (Maybe Text)

The device's IP address.

ethernetStatus_connectionStatus :: Lens' EthernetStatus (Maybe NetworkConnectionStatus)

The device's connection status.

ethernetStatus_hwAddress :: Lens' EthernetStatus (Maybe Text)

The device's physical address.

Job

data Job

A job for a device.

See: newJob smart constructor.

Constructors

Job'

Fields

jobId :: Maybe Text

The job's ID.

deviceId :: Maybe Text

The target device's ID.

newJob :: Job

Create a value of Job with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:Job', job_jobId - The job's ID.

$sel:deviceId:Job', job_deviceId - The target device's ID.

job_jobId :: Lens' Job (Maybe Text)

The job's ID.

job_deviceId :: Lens' Job (Maybe Text)

The target device's ID.

JobResourceTags

data JobResourceTags

Tags for a job.

See: newJobResourceTags smart constructor.

Constructors

JobResourceTags'

Fields

resourceType :: JobResourceType

The job's type.

tags :: HashMap Text Text

The job's tags.

newJobResourceTags

Arguments

:: JobResourceType
$sel:resourceType:JobResourceTags'

-> JobResourceTags

Create a value of JobResourceTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:JobResourceTags', jobResourceTags_resourceType - The job's type.

$sel:tags:JobResourceTags', jobResourceTags_tags - The job's tags.

jobResourceTags_resourceType :: Lens' JobResourceTags JobResourceType

The job's type.

jobResourceTags_tags :: Lens' JobResourceTags (HashMap Text Text)

The job's tags.

ManifestOverridesPayload

data ManifestOverridesPayload

Parameter overrides for an application instance. This is a JSON document that has a single key (PayloadData) where the value is an escaped string representation of the overrides document.

See: newManifestOverridesPayload smart constructor.

Constructors

ManifestOverridesPayload'

Fields

payloadData :: Maybe Text

The overrides document.

newManifestOverridesPayload :: ManifestOverridesPayload

Create a value of ManifestOverridesPayload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:payloadData:ManifestOverridesPayload', manifestOverridesPayload_payloadData - The overrides document.

manifestOverridesPayload_payloadData :: Lens' ManifestOverridesPayload (Maybe Text)

The overrides document.

ManifestPayload

data ManifestPayload

A application verion's manifest file. This is a JSON document that has a single key (PayloadData) where the value is an escaped string representation of the application manifest (graph.json). This file is located in the graphs folder in your application source.

See: newManifestPayload smart constructor.

Constructors

ManifestPayload'

Fields

payloadData :: Maybe Text

The application manifest.

newManifestPayload :: ManifestPayload

Create a value of ManifestPayload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:payloadData:ManifestPayload', manifestPayload_payloadData - The application manifest.

manifestPayload_payloadData :: Lens' ManifestPayload (Maybe Text)

The application manifest.

NetworkPayload

data NetworkPayload

The network configuration for a device.

See: newNetworkPayload smart constructor.

Constructors

NetworkPayload'

Fields

ethernet1 :: Maybe EthernetPayload

Settings for Ethernet port 1.

ethernet0 :: Maybe EthernetPayload

Settings for Ethernet port 0.

newNetworkPayload :: NetworkPayload

Create a value of NetworkPayload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ethernet1:NetworkPayload', networkPayload_ethernet1 - Settings for Ethernet port 1.

$sel:ethernet0:NetworkPayload', networkPayload_ethernet0 - Settings for Ethernet port 0.

networkPayload_ethernet1 :: Lens' NetworkPayload (Maybe EthernetPayload)

Settings for Ethernet port 1.

networkPayload_ethernet0 :: Lens' NetworkPayload (Maybe EthernetPayload)

Settings for Ethernet port 0.

NetworkStatus

data NetworkStatus

The network status of a device.

See: newNetworkStatus smart constructor.

Constructors

NetworkStatus'

Fields

ethernet1Status :: Maybe EthernetStatus

The status of Ethernet port 1.

ethernet0Status :: Maybe EthernetStatus

The status of Ethernet port 0.

newNetworkStatus :: NetworkStatus

Create a value of NetworkStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ethernet1Status:NetworkStatus', networkStatus_ethernet1Status - The status of Ethernet port 1.

$sel:ethernet0Status:NetworkStatus', networkStatus_ethernet0Status - The status of Ethernet port 0.

networkStatus_ethernet1Status :: Lens' NetworkStatus (Maybe EthernetStatus)

The status of Ethernet port 1.

networkStatus_ethernet0Status :: Lens' NetworkStatus (Maybe EthernetStatus)

The status of Ethernet port 0.

Node

data Node

An application node that represents a camera stream, a model, code, or output.

See: newNode smart constructor.

Constructors

Node'

Fields

packageArn :: Maybe Text

The node's ARN.

ownerAccount :: Maybe Text

The account ID of the node's owner.

description :: Maybe Text

The node's description.

nodeId :: Text

The node's ID.

name :: Text

The node's name.

category :: NodeCategory

The node's category.

packageName :: Text

The node's package name.

packageId :: Text

The node's package ID.

packageVersion :: Text

The node's package version.

patchVersion :: Text

The node's patch version.

createdTime :: POSIX

When the node was created.

newNode

Arguments

:: Text
$sel:nodeId:Node'

-> Text
$sel:name:Node'

-> NodeCategory
$sel:category:Node'

-> Text
$sel:packageName:Node'

-> Text
$sel:packageId:Node'

-> Text
$sel:packageVersion:Node'

-> Text
$sel:patchVersion:Node'

-> UTCTime
$sel:createdTime:Node'

-> Node

Create a value of Node with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageArn:Node', node_packageArn - The node's ARN.

$sel:ownerAccount:Node', node_ownerAccount - The account ID of the node's owner.

$sel:description:Node', node_description - The node's description.

$sel:nodeId:Node', node_nodeId - The node's ID.

$sel:name:Node', node_name - The node's name.

$sel:category:Node', node_category - The node's category.

$sel:packageName:Node', node_packageName - The node's package name.

$sel:packageId:Node', node_packageId - The node's package ID.

$sel:packageVersion:Node', node_packageVersion - The node's package version.

$sel:patchVersion:Node', node_patchVersion - The node's patch version.

$sel:createdTime:Node', node_createdTime - When the node was created.

node_packageArn :: Lens' Node (Maybe Text)

The node's ARN.

node_ownerAccount :: Lens' Node (Maybe Text)

The account ID of the node's owner.

node_description :: Lens' Node (Maybe Text)

The node's description.

node_nodeId :: Lens' Node Text

The node's ID.

node_name :: Lens' Node Text

The node's name.

node_category :: Lens' Node NodeCategory

The node's category.

node_packageName :: Lens' Node Text

The node's package name.

node_packageId :: Lens' Node Text

The node's package ID.

node_packageVersion :: Lens' Node Text

The node's package version.

node_patchVersion :: Lens' Node Text

The node's patch version.

node_createdTime :: Lens' Node UTCTime

When the node was created.

NodeFromTemplateJob

data NodeFromTemplateJob

A job to create a camera stream node.

See: newNodeFromTemplateJob smart constructor.

Constructors

NodeFromTemplateJob'

Fields

status :: Maybe NodeFromTemplateJobStatus

The job's status.

jobId :: Maybe Text

The job's ID.

createdTime :: Maybe POSIX

When the job was created.

templateType :: Maybe TemplateType

The job's template type.

nodeName :: Maybe Text

The node's name.

statusMessage :: Maybe Text

The job's status message.

newNodeFromTemplateJob :: NodeFromTemplateJob

Create a value of NodeFromTemplateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:NodeFromTemplateJob', nodeFromTemplateJob_status - The job's status.

$sel:jobId:NodeFromTemplateJob', nodeFromTemplateJob_jobId - The job's ID.

$sel:createdTime:NodeFromTemplateJob', nodeFromTemplateJob_createdTime - When the job was created.

$sel:templateType:NodeFromTemplateJob', nodeFromTemplateJob_templateType - The job's template type.

$sel:nodeName:NodeFromTemplateJob', nodeFromTemplateJob_nodeName - The node's name.

$sel:statusMessage:NodeFromTemplateJob', nodeFromTemplateJob_statusMessage - The job's status message.

nodeFromTemplateJob_status :: Lens' NodeFromTemplateJob (Maybe NodeFromTemplateJobStatus)

The job's status.

nodeFromTemplateJob_jobId :: Lens' NodeFromTemplateJob (Maybe Text)

The job's ID.

nodeFromTemplateJob_createdTime :: Lens' NodeFromTemplateJob (Maybe UTCTime)

When the job was created.

nodeFromTemplateJob_templateType :: Lens' NodeFromTemplateJob (Maybe TemplateType)

The job's template type.

nodeFromTemplateJob_nodeName :: Lens' NodeFromTemplateJob (Maybe Text)

The node's name.

nodeFromTemplateJob_statusMessage :: Lens' NodeFromTemplateJob (Maybe Text)

The job's status message.

NodeInputPort

data NodeInputPort

A node input port.

See: newNodeInputPort smart constructor.

Constructors

NodeInputPort'

Fields

maxConnections :: Maybe Int

The input port's max connections.

name :: Maybe Text

The input port's name.

defaultValue :: Maybe Text

The input port's default value.

type' :: Maybe PortType

The input port's type.

description :: Maybe Text

The input port's description.

newNodeInputPort :: NodeInputPort

Create a value of NodeInputPort with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxConnections:NodeInputPort', nodeInputPort_maxConnections - The input port's max connections.

$sel:name:NodeInputPort', nodeInputPort_name - The input port's name.

$sel:defaultValue:NodeInputPort', nodeInputPort_defaultValue - The input port's default value.

$sel:type':NodeInputPort', nodeInputPort_type - The input port's type.

$sel:description:NodeInputPort', nodeInputPort_description - The input port's description.

nodeInputPort_maxConnections :: Lens' NodeInputPort (Maybe Int)

The input port's max connections.

nodeInputPort_name :: Lens' NodeInputPort (Maybe Text)

The input port's name.

nodeInputPort_defaultValue :: Lens' NodeInputPort (Maybe Text)

The input port's default value.

nodeInputPort_type :: Lens' NodeInputPort (Maybe PortType)

The input port's type.

nodeInputPort_description :: Lens' NodeInputPort (Maybe Text)

The input port's description.

NodeInstance

data NodeInstance

A node instance.

See: newNodeInstance smart constructor.

Constructors

NodeInstance'

Fields

packageName :: Maybe Text

The instance's package name.

packageVersion :: Maybe Text

The instance's package version.

packagePatchVersion :: Maybe Text

The instance's package patch version.

nodeName :: Maybe Text

The instance's name.

nodeId :: Maybe Text

The node's ID.

nodeInstanceId :: Text

The instance's ID.

currentStatus :: NodeInstanceStatus

The instance's current status.

newNodeInstance

Arguments

:: Text
$sel:nodeInstanceId:NodeInstance'

-> NodeInstanceStatus
$sel:currentStatus:NodeInstance'

-> NodeInstance

Create a value of NodeInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageName:NodeInstance', nodeInstance_packageName - The instance's package name.

$sel:packageVersion:NodeInstance', nodeInstance_packageVersion - The instance's package version.

$sel:packagePatchVersion:NodeInstance', nodeInstance_packagePatchVersion - The instance's package patch version.

$sel:nodeName:NodeInstance', nodeInstance_nodeName - The instance's name.

$sel:nodeId:NodeInstance', nodeInstance_nodeId - The node's ID.

$sel:nodeInstanceId:NodeInstance', nodeInstance_nodeInstanceId - The instance's ID.

$sel:currentStatus:NodeInstance', nodeInstance_currentStatus - The instance's current status.

nodeInstance_packageName :: Lens' NodeInstance (Maybe Text)

The instance's package name.

nodeInstance_packageVersion :: Lens' NodeInstance (Maybe Text)

The instance's package version.

nodeInstance_packagePatchVersion :: Lens' NodeInstance (Maybe Text)

The instance's package patch version.

nodeInstance_nodeName :: Lens' NodeInstance (Maybe Text)

The instance's name.

nodeInstance_nodeId :: Lens' NodeInstance (Maybe Text)

The node's ID.

nodeInstance_nodeInstanceId :: Lens' NodeInstance Text

The instance's ID.

nodeInstance_currentStatus :: Lens' NodeInstance NodeInstanceStatus

The instance's current status.

NodeInterface

data NodeInterface

A node interface.

See: newNodeInterface smart constructor.

Constructors

NodeInterface'

Fields

inputs :: [NodeInputPort]

The node interface's inputs.

outputs :: [NodeOutputPort]

The node interface's outputs.

newNodeInterface :: NodeInterface

Create a value of NodeInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputs:NodeInterface', nodeInterface_inputs - The node interface's inputs.

$sel:outputs:NodeInterface', nodeInterface_outputs - The node interface's outputs.

nodeInterface_inputs :: Lens' NodeInterface [NodeInputPort]

The node interface's inputs.

nodeInterface_outputs :: Lens' NodeInterface [NodeOutputPort]

The node interface's outputs.

NodeOutputPort

data NodeOutputPort

A node output port.

See: newNodeOutputPort smart constructor.

Constructors

NodeOutputPort'

Fields

name :: Maybe Text

The output port's name.

type' :: Maybe PortType

The output port's type.

description :: Maybe Text

The output port's description.

newNodeOutputPort :: NodeOutputPort

Create a value of NodeOutputPort with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:NodeOutputPort', nodeOutputPort_name - The output port's name.

$sel:type':NodeOutputPort', nodeOutputPort_type - The output port's type.

$sel:description:NodeOutputPort', nodeOutputPort_description - The output port's description.

nodeOutputPort_name :: Lens' NodeOutputPort (Maybe Text)

The output port's name.

nodeOutputPort_type :: Lens' NodeOutputPort (Maybe PortType)

The output port's type.

nodeOutputPort_description :: Lens' NodeOutputPort (Maybe Text)

The output port's description.

OTAJobConfig

data OTAJobConfig

An over-the-air update (OTA) job configuration.

See: newOTAJobConfig smart constructor.

Constructors

OTAJobConfig'

Fields

imageVersion :: Text

The target version of the device software.

newOTAJobConfig

Arguments

:: Text
$sel:imageVersion:OTAJobConfig'

-> OTAJobConfig

Create a value of OTAJobConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageVersion:OTAJobConfig', oTAJobConfig_imageVersion - The target version of the device software.

oTAJobConfig_imageVersion :: Lens' OTAJobConfig Text

The target version of the device software.

OutPutS3Location

data OutPutS3Location

The location of an output object in Amazon S3.

See: newOutPutS3Location smart constructor.

Constructors

OutPutS3Location'

Fields

bucketName :: Text

The object's bucket.

objectKey :: Text

The object's key.

newOutPutS3Location

Arguments

:: Text
$sel:bucketName:OutPutS3Location'

-> Text
$sel:objectKey:OutPutS3Location'

-> OutPutS3Location

Create a value of OutPutS3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketName:OutPutS3Location', outPutS3Location_bucketName - The object's bucket.

$sel:objectKey:OutPutS3Location', outPutS3Location_objectKey - The object's key.

outPutS3Location_bucketName :: Lens' OutPutS3Location Text

The object's bucket.

outPutS3Location_objectKey :: Lens' OutPutS3Location Text

The object's key.

PackageImportJob

data PackageImportJob

A job to import a package version.

See: newPackageImportJob smart constructor.

Constructors

PackageImportJob'

Fields

status :: Maybe PackageImportJobStatus

The job's status.

jobType :: Maybe PackageImportJobType

The job's type.

lastUpdatedTime :: Maybe POSIX

When the job was updated.

jobId :: Maybe Text

The job's ID.

createdTime :: Maybe POSIX

When the job was created.

statusMessage :: Maybe Text

The job's status message.

newPackageImportJob :: PackageImportJob

Create a value of PackageImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:PackageImportJob', packageImportJob_status - The job's status.

$sel:jobType:PackageImportJob', packageImportJob_jobType - The job's type.

$sel:lastUpdatedTime:PackageImportJob', packageImportJob_lastUpdatedTime - When the job was updated.

$sel:jobId:PackageImportJob', packageImportJob_jobId - The job's ID.

$sel:createdTime:PackageImportJob', packageImportJob_createdTime - When the job was created.

$sel:statusMessage:PackageImportJob', packageImportJob_statusMessage - The job's status message.

packageImportJob_status :: Lens' PackageImportJob (Maybe PackageImportJobStatus)

The job's status.

packageImportJob_jobType :: Lens' PackageImportJob (Maybe PackageImportJobType)

The job's type.

packageImportJob_lastUpdatedTime :: Lens' PackageImportJob (Maybe UTCTime)

When the job was updated.

packageImportJob_jobId :: Lens' PackageImportJob (Maybe Text)

The job's ID.

packageImportJob_createdTime :: Lens' PackageImportJob (Maybe UTCTime)

When the job was created.

packageImportJob_statusMessage :: Lens' PackageImportJob (Maybe Text)

The job's status message.

PackageImportJobInputConfig

data PackageImportJobInputConfig

A configuration for a package import job.

See: newPackageImportJobInputConfig smart constructor.

Constructors

PackageImportJobInputConfig'

Fields

packageVersionInputConfig :: Maybe PackageVersionInputConfig

The package version's input configuration.

newPackageImportJobInputConfig :: PackageImportJobInputConfig

Create a value of PackageImportJobInputConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageVersionInputConfig:PackageImportJobInputConfig', packageImportJobInputConfig_packageVersionInputConfig - The package version's input configuration.

packageImportJobInputConfig_packageVersionInputConfig :: Lens' PackageImportJobInputConfig (Maybe PackageVersionInputConfig)

The package version's input configuration.

PackageImportJobOutput

data PackageImportJobOutput

Results of a package import job.

See: newPackageImportJobOutput smart constructor.

Constructors

PackageImportJobOutput'

Fields

packageId :: Text

The package's ID.

packageVersion :: Text

The package's version.

patchVersion :: Text

The package's patch version.

outputS3Location :: OutPutS3Location

The package's output location.

newPackageImportJobOutput

Arguments

:: Text
$sel:packageId:PackageImportJobOutput'

-> Text
$sel:packageVersion:PackageImportJobOutput'

-> Text
$sel:patchVersion:PackageImportJobOutput'

-> OutPutS3Location
$sel:outputS3Location:PackageImportJobOutput'

-> PackageImportJobOutput

Create a value of PackageImportJobOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageId:PackageImportJobOutput', packageImportJobOutput_packageId - The package's ID.

$sel:packageVersion:PackageImportJobOutput', packageImportJobOutput_packageVersion - The package's version.

$sel:patchVersion:PackageImportJobOutput', packageImportJobOutput_patchVersion - The package's patch version.

$sel:outputS3Location:PackageImportJobOutput', packageImportJobOutput_outputS3Location - The package's output location.

packageImportJobOutput_packageId :: Lens' PackageImportJobOutput Text

The package's ID.

packageImportJobOutput_packageVersion :: Lens' PackageImportJobOutput Text

The package's version.

packageImportJobOutput_patchVersion :: Lens' PackageImportJobOutput Text

The package's patch version.

packageImportJobOutput_outputS3Location :: Lens' PackageImportJobOutput OutPutS3Location

The package's output location.

PackageImportJobOutputConfig

data PackageImportJobOutputConfig

An output configuration for a package import job.

See: newPackageImportJobOutputConfig smart constructor.

Constructors

PackageImportJobOutputConfig'

Fields

packageVersionOutputConfig :: Maybe PackageVersionOutputConfig

The package version's output configuration.

newPackageImportJobOutputConfig :: PackageImportJobOutputConfig

Create a value of PackageImportJobOutputConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageVersionOutputConfig:PackageImportJobOutputConfig', packageImportJobOutputConfig_packageVersionOutputConfig - The package version's output configuration.

packageImportJobOutputConfig_packageVersionOutputConfig :: Lens' PackageImportJobOutputConfig (Maybe PackageVersionOutputConfig)

The package version's output configuration.

PackageListItem

data PackageListItem

A package summary.

See: newPackageListItem smart constructor.

Constructors

PackageListItem'

Fields

packageId :: Maybe Text

The package's ID.

arn :: Maybe Text

The package's ARN.

createdTime :: Maybe POSIX

When the package was created.

packageName :: Maybe Text

The package's name.

tags :: Maybe (HashMap Text Text)

The package's tags.

newPackageListItem :: PackageListItem

Create a value of PackageListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageId:PackageListItem', packageListItem_packageId - The package's ID.

$sel:arn:PackageListItem', packageListItem_arn - The package's ARN.

$sel:createdTime:PackageListItem', packageListItem_createdTime - When the package was created.

$sel:packageName:PackageListItem', packageListItem_packageName - The package's name.

$sel:tags:PackageListItem', packageListItem_tags - The package's tags.

packageListItem_packageId :: Lens' PackageListItem (Maybe Text)

The package's ID.

packageListItem_arn :: Lens' PackageListItem (Maybe Text)

The package's ARN.

packageListItem_createdTime :: Lens' PackageListItem (Maybe UTCTime)

When the package was created.

packageListItem_packageName :: Lens' PackageListItem (Maybe Text)

The package's name.

packageListItem_tags :: Lens' PackageListItem (Maybe (HashMap Text Text))

The package's tags.

PackageObject

data PackageObject

A package object.

See: newPackageObject smart constructor.

Constructors

PackageObject'

Fields

name :: Text

The object's name.

packageVersion :: Text

The object's package version.

patchVersion :: Text

The object's patch version.

newPackageObject

Arguments

:: Text
$sel:name:PackageObject'

-> Text
$sel:packageVersion:PackageObject'

-> Text
$sel:patchVersion:PackageObject'

-> PackageObject

Create a value of PackageObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:PackageObject', packageObject_name - The object's name.

$sel:packageVersion:PackageObject', packageObject_packageVersion - The object's package version.

$sel:patchVersion:PackageObject', packageObject_patchVersion - The object's patch version.

packageObject_name :: Lens' PackageObject Text

The object's name.

packageObject_packageVersion :: Lens' PackageObject Text

The object's package version.

packageObject_patchVersion :: Lens' PackageObject Text

The object's patch version.

PackageVersionInputConfig

data PackageVersionInputConfig

A package version input configuration.

See: newPackageVersionInputConfig smart constructor.

Constructors

PackageVersionInputConfig'

Fields

s3Location :: S3Location

A location in Amazon S3.

newPackageVersionInputConfig

Arguments

:: S3Location
$sel:s3Location:PackageVersionInputConfig'

-> PackageVersionInputConfig

Create a value of PackageVersionInputConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Location:PackageVersionInputConfig', packageVersionInputConfig_s3Location - A location in Amazon S3.

packageVersionInputConfig_s3Location :: Lens' PackageVersionInputConfig S3Location

A location in Amazon S3.

PackageVersionOutputConfig

data PackageVersionOutputConfig

A package version output configuration.

See: newPackageVersionOutputConfig smart constructor.

Constructors

PackageVersionOutputConfig'

Fields

markLatest :: Maybe Bool

Indicates that the version is recommended for all users.

packageName :: Text

The output's package name.

packageVersion :: Text

The output's package version.

newPackageVersionOutputConfig

Arguments

:: Text
$sel:packageName:PackageVersionOutputConfig'

-> Text
$sel:packageVersion:PackageVersionOutputConfig'

-> PackageVersionOutputConfig

Create a value of PackageVersionOutputConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:markLatest:PackageVersionOutputConfig', packageVersionOutputConfig_markLatest - Indicates that the version is recommended for all users.

$sel:packageName:PackageVersionOutputConfig', packageVersionOutputConfig_packageName - The output's package name.

$sel:packageVersion:PackageVersionOutputConfig', packageVersionOutputConfig_packageVersion - The output's package version.

packageVersionOutputConfig_markLatest :: Lens' PackageVersionOutputConfig (Maybe Bool)

Indicates that the version is recommended for all users.

packageVersionOutputConfig_packageName :: Lens' PackageVersionOutputConfig Text

The output's package name.

packageVersionOutputConfig_packageVersion :: Lens' PackageVersionOutputConfig Text

The output's package version.

S3Location

data S3Location

A location in Amazon S3.

See: newS3Location smart constructor.

Constructors

S3Location'

Fields

region :: Maybe Text

The bucket's Region.

bucketName :: Text

A bucket name.

objectKey :: Text

An object key.

newS3Location

Arguments

:: Text
$sel:bucketName:S3Location'

-> Text
$sel:objectKey:S3Location'

-> S3Location

Create a value of S3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:region:S3Location', s3Location_region - The bucket's Region.

$sel:bucketName:S3Location', s3Location_bucketName - A bucket name.

$sel:objectKey:S3Location', s3Location_objectKey - An object key.

s3Location_region :: Lens' S3Location (Maybe Text)

The bucket's Region.

s3Location_bucketName :: Lens' S3Location Text

A bucket name.

s3Location_objectKey :: Lens' S3Location Text

An object key.

StaticIpConnectionInfo

data StaticIpConnectionInfo

A static IP configuration.

See: newStaticIpConnectionInfo smart constructor.

Constructors

StaticIpConnectionInfo'

Fields

ipAddress :: Text

The connection's IP address.

mask :: Text

The connection's DNS mask.

dns :: [Text]

The connection's DNS address.

defaultGateway :: Text

The connection's default gateway.

newStaticIpConnectionInfo

Arguments

:: Text
$sel:ipAddress:StaticIpConnectionInfo'

-> Text
$sel:mask:StaticIpConnectionInfo'

-> Text
$sel:defaultGateway:StaticIpConnectionInfo'

-> StaticIpConnectionInfo

Create a value of StaticIpConnectionInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddress:StaticIpConnectionInfo', staticIpConnectionInfo_ipAddress - The connection's IP address.

$sel:mask:StaticIpConnectionInfo', staticIpConnectionInfo_mask - The connection's DNS mask.

$sel:dns:StaticIpConnectionInfo', staticIpConnectionInfo_dns - The connection's DNS address.

$sel:defaultGateway:StaticIpConnectionInfo', staticIpConnectionInfo_defaultGateway - The connection's default gateway.

staticIpConnectionInfo_ipAddress :: Lens' StaticIpConnectionInfo Text

The connection's IP address.

staticIpConnectionInfo_mask :: Lens' StaticIpConnectionInfo Text

The connection's DNS mask.

staticIpConnectionInfo_dns :: Lens' StaticIpConnectionInfo [Text]

The connection's DNS address.

staticIpConnectionInfo_defaultGateway :: Lens' StaticIpConnectionInfo Text

The connection's default gateway.

StorageLocation

data StorageLocation

A storage location.

See: newStorageLocation smart constructor.

Constructors

StorageLocation'

Fields

bucket :: Text

The location's bucket.

repoPrefixLocation :: Text

The location's repo prefix.

generatedPrefixLocation :: Text

The location's generated prefix.

binaryPrefixLocation :: Text

The location's binary prefix.

manifestPrefixLocation :: Text

The location's manifest prefix.

newStorageLocation

Arguments

:: Text
$sel:bucket:StorageLocation'

-> Text
$sel:repoPrefixLocation:StorageLocation'

-> Text
$sel:generatedPrefixLocation:StorageLocation'

-> Text
$sel:binaryPrefixLocation:StorageLocation'

-> Text
$sel:manifestPrefixLocation:StorageLocation'

-> StorageLocation

Create a value of StorageLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucket:StorageLocation', storageLocation_bucket - The location's bucket.

$sel:repoPrefixLocation:StorageLocation', storageLocation_repoPrefixLocation - The location's repo prefix.

$sel:generatedPrefixLocation:StorageLocation', storageLocation_generatedPrefixLocation - The location's generated prefix.

$sel:binaryPrefixLocation:StorageLocation', storageLocation_binaryPrefixLocation - The location's binary prefix.

$sel:manifestPrefixLocation:StorageLocation', storageLocation_manifestPrefixLocation - The location's manifest prefix.

storageLocation_bucket :: Lens' StorageLocation Text

The location's bucket.

storageLocation_repoPrefixLocation :: Lens' StorageLocation Text

The location's repo prefix.

storageLocation_generatedPrefixLocation :: Lens' StorageLocation Text

The location's generated prefix.

storageLocation_binaryPrefixLocation :: Lens' StorageLocation Text

The location's binary prefix.

storageLocation_manifestPrefixLocation :: Lens' StorageLocation Text

The location's manifest prefix.