Amazonka.Personalize.CreateCampaign

Description

Creates a campaign by deploying a solution version. When a client calls the GetRecommendations and GetPersonalizedRanking APIs, a campaign is specified in the request.

Minimum Provisioned TPS and Auto-Scaling

A transaction is a single GetRecommendations or GetPersonalizedRanking call. Transactions per second (TPS) is the throughput and unit of billing for Amazon Personalize. The minimum provisioned TPS (minProvisionedTPS) specifies the baseline throughput provisioned by Amazon Personalize, and thus, the minimum billing charge.

If your TPS increases beyond minProvisionedTPS, Amazon Personalize auto-scales the provisioned capacity up and down, but never below minProvisionedTPS. There's a short time delay while the capacity is increased that might cause loss of transactions.

The actual TPS used is calculated as the average requests/second within a 5-minute window. You pay for maximum of either the minimum provisioned TPS or the actual TPS. We recommend starting with a low minProvisionedTPS, track your usage using Amazon CloudWatch metrics, and then increase the minProvisionedTPS as necessary.

Status

A campaign can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING > DELETE IN_PROGRESS

To get the campaign status, call DescribeCampaign.

Wait until the status of the campaign is ACTIVE before asking the campaign for recommendations.

Related APIs

ListCampaigns

DescribeCampaign

UpdateCampaign

DeleteCampaign

Creating a Request

data CreateCampaign

See: newCreateCampaign smart constructor.

Constructors

CreateCampaign'

Fields

campaignConfig :: Maybe CampaignConfig

The configuration details of a campaign.

minProvisionedTPS :: Maybe Natural

Specifies the requested minimum provisioned transactions (recommendations) per second that Amazon Personalize will support.

name :: Text

A name for the new campaign. The campaign name must be unique within your account.

solutionVersionArn :: Text

The Amazon Resource Name (ARN) of the solution version to deploy.

newCreateCampaign

Arguments

:: Text
$sel:name:CreateCampaign'

-> Text
$sel:solutionVersionArn:CreateCampaign'

-> CreateCampaign

Create a value of CreateCampaign with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:campaignConfig:CreateCampaign', createCampaign_campaignConfig - The configuration details of a campaign.

$sel:minProvisionedTPS:CreateCampaign', createCampaign_minProvisionedTPS - Specifies the requested minimum provisioned transactions (recommendations) per second that Amazon Personalize will support.

$sel:name:CreateCampaign', createCampaign_name - A name for the new campaign. The campaign name must be unique within your account.

$sel:solutionVersionArn:CreateCampaign', createCampaign_solutionVersionArn - The Amazon Resource Name (ARN) of the solution version to deploy.

Request Lenses

createCampaign_campaignConfig :: Lens' CreateCampaign (Maybe CampaignConfig)

The configuration details of a campaign.

createCampaign_minProvisionedTPS :: Lens' CreateCampaign (Maybe Natural)

Specifies the requested minimum provisioned transactions (recommendations) per second that Amazon Personalize will support.

createCampaign_name :: Lens' CreateCampaign Text

A name for the new campaign. The campaign name must be unique within your account.

createCampaign_solutionVersionArn :: Lens' CreateCampaign Text

The Amazon Resource Name (ARN) of the solution version to deploy.

Destructuring the Response

data CreateCampaignResponse

See: newCreateCampaignResponse smart constructor.

Constructors

CreateCampaignResponse'

Fields

campaignArn :: Maybe Text

The Amazon Resource Name (ARN) of the campaign.

httpStatus :: Int

The response's http status code.

newCreateCampaignResponse

Arguments

:: Int
$sel:httpStatus:CreateCampaignResponse'

-> CreateCampaignResponse

Create a value of CreateCampaignResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:campaignArn:CreateCampaignResponse', createCampaignResponse_campaignArn - The Amazon Resource Name (ARN) of the campaign.

$sel:httpStatus:CreateCampaignResponse', createCampaignResponse_httpStatus - The response's http status code.

Response Lenses

createCampaignResponse_campaignArn :: Lens' CreateCampaignResponse (Maybe Text)

The Amazon Resource Name (ARN) of the campaign.

createCampaignResponse_httpStatus :: Lens' CreateCampaignResponse Int

The response's http status code.