Amazonka.Personalize.CreateDatasetImportJob

Description

Creates a job that imports training data from your data source (an Amazon S3 bucket) to an Amazon Personalize dataset. To allow Amazon Personalize to import the training data, you must specify an IAM service role that has permission to read from the data source, as Amazon Personalize makes a copy of your data and processes it internally. For information on granting access to your Amazon S3 bucket, see Giving Amazon Personalize Access to Amazon S3 Resources.

The dataset import job replaces any existing data in the dataset that you imported in bulk.

Status

A dataset import job can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

To get the status of the import job, call DescribeDatasetImportJob, providing the Amazon Resource Name (ARN) of the dataset import job. The dataset import is complete when the status shows as ACTIVE. If the status shows as CREATE FAILED, the response includes a failureReason key, which describes why the job failed.

Importing takes time. You must wait until the status shows as ACTIVE before training a model using the dataset.

Related APIs

ListDatasetImportJobs

DescribeDatasetImportJob

Creating a Request

data CreateDatasetImportJob

See: newCreateDatasetImportJob smart constructor.

Constructors

CreateDatasetImportJob'

Fields

jobName :: Text

The name for the dataset import job.

datasetArn :: Text

The ARN of the dataset that receives the imported data.

dataSource :: DataSource

The Amazon S3 bucket that contains the training data to import.

roleArn :: Text

The ARN of the IAM role that has permissions to read from the Amazon S3 data source.

newCreateDatasetImportJob

Arguments

:: Text
$sel:jobName:CreateDatasetImportJob'

-> Text
$sel:datasetArn:CreateDatasetImportJob'

-> DataSource
$sel:dataSource:CreateDatasetImportJob'

-> Text
$sel:roleArn:CreateDatasetImportJob'

-> CreateDatasetImportJob

Create a value of CreateDatasetImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobName:CreateDatasetImportJob', createDatasetImportJob_jobName - The name for the dataset import job.

$sel:datasetArn:CreateDatasetImportJob', createDatasetImportJob_datasetArn - The ARN of the dataset that receives the imported data.

$sel:dataSource:CreateDatasetImportJob', createDatasetImportJob_dataSource - The Amazon S3 bucket that contains the training data to import.

$sel:roleArn:CreateDatasetImportJob', createDatasetImportJob_roleArn - The ARN of the IAM role that has permissions to read from the Amazon S3 data source.

Request Lenses

createDatasetImportJob_jobName :: Lens' CreateDatasetImportJob Text

The name for the dataset import job.

createDatasetImportJob_datasetArn :: Lens' CreateDatasetImportJob Text

The ARN of the dataset that receives the imported data.

createDatasetImportJob_dataSource :: Lens' CreateDatasetImportJob DataSource

The Amazon S3 bucket that contains the training data to import.

createDatasetImportJob_roleArn :: Lens' CreateDatasetImportJob Text

The ARN of the IAM role that has permissions to read from the Amazon S3 data source.

Destructuring the Response

data CreateDatasetImportJobResponse

See: newCreateDatasetImportJobResponse smart constructor.

Constructors

CreateDatasetImportJobResponse'

Fields

datasetImportJobArn :: Maybe Text

The ARN of the dataset import job.

httpStatus :: Int

The response's http status code.

newCreateDatasetImportJobResponse

Arguments

:: Int
$sel:httpStatus:CreateDatasetImportJobResponse'

-> CreateDatasetImportJobResponse

Create a value of CreateDatasetImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetImportJobArn:CreateDatasetImportJobResponse', createDatasetImportJobResponse_datasetImportJobArn - The ARN of the dataset import job.

$sel:httpStatus:CreateDatasetImportJobResponse', createDatasetImportJobResponse_httpStatus - The response's http status code.

Response Lenses

createDatasetImportJobResponse_datasetImportJobArn :: Lens' CreateDatasetImportJobResponse (Maybe Text)

The ARN of the dataset import job.

createDatasetImportJobResponse_httpStatus :: Lens' CreateDatasetImportJobResponse Int

The response's http status code.