Amazonka.Personalize.DescribeDatasetExportJob

Description

Describes the dataset export job created by CreateDatasetExportJob, including the export job status.

Creating a Request

data DescribeDatasetExportJob

See: newDescribeDatasetExportJob smart constructor.

Constructors

DescribeDatasetExportJob'

Fields

datasetExportJobArn :: Text

The Amazon Resource Name (ARN) of the dataset export job to describe.

newDescribeDatasetExportJob

Arguments

:: Text
$sel:datasetExportJobArn:DescribeDatasetExportJob'

-> DescribeDatasetExportJob

Create a value of DescribeDatasetExportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetExportJobArn:DescribeDatasetExportJob', describeDatasetExportJob_datasetExportJobArn - The Amazon Resource Name (ARN) of the dataset export job to describe.

Request Lenses

describeDatasetExportJob_datasetExportJobArn :: Lens' DescribeDatasetExportJob Text

The Amazon Resource Name (ARN) of the dataset export job to describe.

Destructuring the Response

data DescribeDatasetExportJobResponse

See: newDescribeDatasetExportJobResponse smart constructor.

Constructors

DescribeDatasetExportJobResponse'

Fields

datasetExportJob :: Maybe DatasetExportJob

Information about the dataset export job, including the status.

The status is one of the following values:

CREATE PENDING

CREATE IN_PROGRESS

ACTIVE

CREATE FAILED

httpStatus :: Int

The response's http status code.

newDescribeDatasetExportJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeDatasetExportJobResponse'

-> DescribeDatasetExportJobResponse

Create a value of DescribeDatasetExportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetExportJob:DescribeDatasetExportJobResponse', describeDatasetExportJobResponse_datasetExportJob - Information about the dataset export job, including the status.

The status is one of the following values:

CREATE PENDING

CREATE IN_PROGRESS

ACTIVE

CREATE FAILED

$sel:httpStatus:DescribeDatasetExportJobResponse', describeDatasetExportJobResponse_httpStatus - The response's http status code.

Response Lenses

describeDatasetExportJobResponse_datasetExportJob :: Lens' DescribeDatasetExportJobResponse (Maybe DatasetExportJob)

Information about the dataset export job, including the status.

The status is one of the following values:

CREATE PENDING

CREATE IN_PROGRESS

ACTIVE

CREATE FAILED

describeDatasetExportJobResponse_httpStatus :: Lens' DescribeDatasetExportJobResponse Int

The response's http status code.