Amazonka.Personalize.ListEventTrackers

Description

Returns the list of event trackers associated with the account. The response provides the properties for each event tracker, including the Amazon Resource Name (ARN) and tracking ID. For more information on event trackers, see CreateEventTracker.

This operation returns paginated results.

Creating a Request

data ListEventTrackers

See: newListEventTrackers smart constructor.

Constructors

ListEventTrackers'

Fields

nextToken :: Maybe Text

A token returned from the previous call to ListEventTrackers for getting the next set of event trackers (if they exist).

datasetGroupArn :: Maybe Text

The ARN of a dataset group used to filter the response.

maxResults :: Maybe Natural

The maximum number of event trackers to return.

newListEventTrackers :: ListEventTrackers

Create a value of ListEventTrackers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEventTrackers', listEventTrackers_nextToken - A token returned from the previous call to ListEventTrackers for getting the next set of event trackers (if they exist).

$sel:datasetGroupArn:ListEventTrackers', listEventTrackers_datasetGroupArn - The ARN of a dataset group used to filter the response.

$sel:maxResults:ListEventTrackers', listEventTrackers_maxResults - The maximum number of event trackers to return.

Request Lenses

listEventTrackers_nextToken :: Lens' ListEventTrackers (Maybe Text)

A token returned from the previous call to ListEventTrackers for getting the next set of event trackers (if they exist).

listEventTrackers_datasetGroupArn :: Lens' ListEventTrackers (Maybe Text)

The ARN of a dataset group used to filter the response.

listEventTrackers_maxResults :: Lens' ListEventTrackers (Maybe Natural)

The maximum number of event trackers to return.

Destructuring the Response

data ListEventTrackersResponse

See: newListEventTrackersResponse smart constructor.

Constructors

ListEventTrackersResponse'

Fields

eventTrackers :: Maybe [EventTrackerSummary]

A list of event trackers.

nextToken :: Maybe Text

A token for getting the next set of event trackers (if they exist).

httpStatus :: Int

The response's http status code.

newListEventTrackersResponse

Arguments

:: Int
$sel:httpStatus:ListEventTrackersResponse'

-> ListEventTrackersResponse

Create a value of ListEventTrackersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventTrackers:ListEventTrackersResponse', listEventTrackersResponse_eventTrackers - A list of event trackers.

$sel:nextToken:ListEventTrackers', listEventTrackersResponse_nextToken - A token for getting the next set of event trackers (if they exist).

$sel:httpStatus:ListEventTrackersResponse', listEventTrackersResponse_httpStatus - The response's http status code.

Response Lenses

listEventTrackersResponse_eventTrackers :: Lens' ListEventTrackersResponse (Maybe [EventTrackerSummary])

A list of event trackers.

listEventTrackersResponse_nextToken :: Lens' ListEventTrackersResponse (Maybe Text)

A token for getting the next set of event trackers (if they exist).

listEventTrackersResponse_httpStatus :: Lens' ListEventTrackersResponse Int

The response's http status code.