Amazonka.Personalize.StopSolutionVersionCreation

Description

Stops creating a solution version that is in a state of CREATE_PENDING or CREATE IN_PROGRESS.

Depending on the current state of the solution version, the solution version state changes as follows:

CREATE_PENDING > CREATE_STOPPED

or

CREATE_IN_PROGRESS > CREATE_STOPPING > CREATE_STOPPED

You are billed for all of the training completed up until you stop the solution version creation. You cannot resume creating a solution version once it has been stopped.

Creating a Request

data StopSolutionVersionCreation

See: newStopSolutionVersionCreation smart constructor.

Constructors

StopSolutionVersionCreation'

Fields

solutionVersionArn :: Text

The Amazon Resource Name (ARN) of the solution version you want to stop creating.

newStopSolutionVersionCreation

Arguments

:: Text
$sel:solutionVersionArn:StopSolutionVersionCreation'

-> StopSolutionVersionCreation

Create a value of StopSolutionVersionCreation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:solutionVersionArn:StopSolutionVersionCreation', stopSolutionVersionCreation_solutionVersionArn - The Amazon Resource Name (ARN) of the solution version you want to stop creating.

Request Lenses

stopSolutionVersionCreation_solutionVersionArn :: Lens' StopSolutionVersionCreation Text

The Amazon Resource Name (ARN) of the solution version you want to stop creating.

Destructuring the Response

data StopSolutionVersionCreationResponse

See: newStopSolutionVersionCreationResponse smart constructor.

Constructors

StopSolutionVersionCreationResponse'

newStopSolutionVersionCreationResponse :: StopSolutionVersionCreationResponse

Create a value of StopSolutionVersionCreationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.