Amazonka.Personalize.Types.BatchInferenceJob

Description

Documentation

data BatchInferenceJob

Contains information on a batch inference job.

See: newBatchInferenceJob smart constructor.

Constructors

BatchInferenceJob'

Fields

failureReason :: Maybe Text

If the batch inference job failed, the reason for the failure.

status :: Maybe Text

The status of the batch inference job. The status is one of the following values:

PENDING

IN PROGRESS

ACTIVE

CREATE FAILED

jobOutput :: Maybe BatchInferenceJobOutput

The Amazon S3 bucket that contains the output data generated by the batch inference job.

jobName :: Maybe Text

The name of the batch inference job.

lastUpdatedDateTime :: Maybe POSIX

The time at which the batch inference job was last updated.

numResults :: Maybe Int

The number of recommendations generated by the batch inference job. This number includes the error messages generated for failed input records.

batchInferenceJobConfig :: Maybe BatchInferenceJobConfig

A string to string map of the configuration details of a batch inference job.

batchInferenceJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the batch inference job.

filterArn :: Maybe Text

The ARN of the filter used on the batch inference job.

creationDateTime :: Maybe POSIX

The time at which the batch inference job was created.

solutionVersionArn :: Maybe Text

The Amazon Resource Name (ARN) of the solution version from which the batch inference job was created.

roleArn :: Maybe Text

The ARN of the Amazon Identity and Access Management (IAM) role that requested the batch inference job.

jobInput :: Maybe BatchInferenceJobInput

The Amazon S3 path that leads to the input data used to generate the batch inference job.

newBatchInferenceJob :: BatchInferenceJob

Create a value of BatchInferenceJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:BatchInferenceJob', batchInferenceJob_failureReason - If the batch inference job failed, the reason for the failure.

$sel:status:BatchInferenceJob', batchInferenceJob_status - The status of the batch inference job. The status is one of the following values:

PENDING

IN PROGRESS

ACTIVE

CREATE FAILED

$sel:jobOutput:BatchInferenceJob', batchInferenceJob_jobOutput - The Amazon S3 bucket that contains the output data generated by the batch inference job.

$sel:jobName:BatchInferenceJob', batchInferenceJob_jobName - The name of the batch inference job.

$sel:lastUpdatedDateTime:BatchInferenceJob', batchInferenceJob_lastUpdatedDateTime - The time at which the batch inference job was last updated.

$sel:numResults:BatchInferenceJob', batchInferenceJob_numResults - The number of recommendations generated by the batch inference job. This number includes the error messages generated for failed input records.

$sel:batchInferenceJobConfig:BatchInferenceJob', batchInferenceJob_batchInferenceJobConfig - A string to string map of the configuration details of a batch inference job.

$sel:batchInferenceJobArn:BatchInferenceJob', batchInferenceJob_batchInferenceJobArn - The Amazon Resource Name (ARN) of the batch inference job.

$sel:filterArn:BatchInferenceJob', batchInferenceJob_filterArn - The ARN of the filter used on the batch inference job.

$sel:creationDateTime:BatchInferenceJob', batchInferenceJob_creationDateTime - The time at which the batch inference job was created.

$sel:solutionVersionArn:BatchInferenceJob', batchInferenceJob_solutionVersionArn - The Amazon Resource Name (ARN) of the solution version from which the batch inference job was created.

$sel:roleArn:BatchInferenceJob', batchInferenceJob_roleArn - The ARN of the Amazon Identity and Access Management (IAM) role that requested the batch inference job.

$sel:jobInput:BatchInferenceJob', batchInferenceJob_jobInput - The Amazon S3 path that leads to the input data used to generate the batch inference job.

batchInferenceJob_failureReason :: Lens' BatchInferenceJob (Maybe Text)

If the batch inference job failed, the reason for the failure.

batchInferenceJob_status :: Lens' BatchInferenceJob (Maybe Text)

The status of the batch inference job. The status is one of the following values:

PENDING

IN PROGRESS

ACTIVE

CREATE FAILED

batchInferenceJob_jobOutput :: Lens' BatchInferenceJob (Maybe BatchInferenceJobOutput)

The Amazon S3 bucket that contains the output data generated by the batch inference job.

batchInferenceJob_jobName :: Lens' BatchInferenceJob (Maybe Text)

The name of the batch inference job.

batchInferenceJob_lastUpdatedDateTime :: Lens' BatchInferenceJob (Maybe UTCTime)

The time at which the batch inference job was last updated.

batchInferenceJob_numResults :: Lens' BatchInferenceJob (Maybe Int)

The number of recommendations generated by the batch inference job. This number includes the error messages generated for failed input records.

batchInferenceJob_batchInferenceJobConfig :: Lens' BatchInferenceJob (Maybe BatchInferenceJobConfig)

A string to string map of the configuration details of a batch inference job.

batchInferenceJob_batchInferenceJobArn :: Lens' BatchInferenceJob (Maybe Text)

The Amazon Resource Name (ARN) of the batch inference job.

batchInferenceJob_filterArn :: Lens' BatchInferenceJob (Maybe Text)

The ARN of the filter used on the batch inference job.

batchInferenceJob_creationDateTime :: Lens' BatchInferenceJob (Maybe UTCTime)

The time at which the batch inference job was created.

batchInferenceJob_solutionVersionArn :: Lens' BatchInferenceJob (Maybe Text)

The Amazon Resource Name (ARN) of the solution version from which the batch inference job was created.

batchInferenceJob_roleArn :: Lens' BatchInferenceJob (Maybe Text)

The ARN of the Amazon Identity and Access Management (IAM) role that requested the batch inference job.

batchInferenceJob_jobInput :: Lens' BatchInferenceJob (Maybe BatchInferenceJobInput)

The Amazon S3 path that leads to the input data used to generate the batch inference job.