Amazonka.Personalize.Types.BatchInferenceJobSummary

Description

Documentation

data BatchInferenceJobSummary

A truncated version of the BatchInferenceJob datatype. The ListBatchInferenceJobs operation returns a list of batch inference job summaries.

See: newBatchInferenceJobSummary smart constructor.

Constructors

BatchInferenceJobSummary'

Fields

failureReason :: Maybe Text

If the batch inference job failed, the reason for the failure.

status :: Maybe Text

The status of the batch inference job. The status is one of the following values:

PENDING

IN PROGRESS

ACTIVE

CREATE FAILED

jobName :: Maybe Text

The name of the batch inference job.

lastUpdatedDateTime :: Maybe POSIX

The time at which the batch inference job was last updated.

batchInferenceJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the batch inference job.

creationDateTime :: Maybe POSIX

The time at which the batch inference job was created.

solutionVersionArn :: Maybe Text

The ARN of the solution version used by the batch inference job.

newBatchInferenceJobSummary :: BatchInferenceJobSummary

Create a value of BatchInferenceJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:BatchInferenceJobSummary', batchInferenceJobSummary_failureReason - If the batch inference job failed, the reason for the failure.

$sel:status:BatchInferenceJobSummary', batchInferenceJobSummary_status - The status of the batch inference job. The status is one of the following values:

PENDING

IN PROGRESS

ACTIVE

CREATE FAILED

$sel:jobName:BatchInferenceJobSummary', batchInferenceJobSummary_jobName - The name of the batch inference job.

$sel:lastUpdatedDateTime:BatchInferenceJobSummary', batchInferenceJobSummary_lastUpdatedDateTime - The time at which the batch inference job was last updated.

$sel:batchInferenceJobArn:BatchInferenceJobSummary', batchInferenceJobSummary_batchInferenceJobArn - The Amazon Resource Name (ARN) of the batch inference job.

$sel:creationDateTime:BatchInferenceJobSummary', batchInferenceJobSummary_creationDateTime - The time at which the batch inference job was created.

$sel:solutionVersionArn:BatchInferenceJobSummary', batchInferenceJobSummary_solutionVersionArn - The ARN of the solution version used by the batch inference job.

batchInferenceJobSummary_failureReason :: Lens' BatchInferenceJobSummary (Maybe Text)

If the batch inference job failed, the reason for the failure.

batchInferenceJobSummary_status :: Lens' BatchInferenceJobSummary (Maybe Text)

The status of the batch inference job. The status is one of the following values:

PENDING

IN PROGRESS

ACTIVE

CREATE FAILED

batchInferenceJobSummary_jobName :: Lens' BatchInferenceJobSummary (Maybe Text)

The name of the batch inference job.

batchInferenceJobSummary_lastUpdatedDateTime :: Lens' BatchInferenceJobSummary (Maybe UTCTime)

The time at which the batch inference job was last updated.

batchInferenceJobSummary_batchInferenceJobArn :: Lens' BatchInferenceJobSummary (Maybe Text)

The Amazon Resource Name (ARN) of the batch inference job.

batchInferenceJobSummary_creationDateTime :: Lens' BatchInferenceJobSummary (Maybe UTCTime)

The time at which the batch inference job was created.

batchInferenceJobSummary_solutionVersionArn :: Lens' BatchInferenceJobSummary (Maybe Text)

The ARN of the solution version used by the batch inference job.