Amazonka.Personalize.Types.Campaign

Description

Documentation

data Campaign

Describes a deployed solution version, otherwise known as a campaign. For more information on campaigns, see CreateCampaign.

See: newCampaign smart constructor.

Constructors

Campaign'

Fields

failureReason :: Maybe Text

If a campaign fails, the reason behind the failure.

status :: Maybe Text

The status of the campaign.

A campaign can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING > DELETE IN_PROGRESS

lastUpdatedDateTime :: Maybe POSIX

The date and time (in Unix format) that the campaign was last updated.

campaignConfig :: Maybe CampaignConfig

The configuration details of a campaign.

latestCampaignUpdate :: Maybe CampaignUpdateSummary

name :: Maybe Text

The name of the campaign.

minProvisionedTPS :: Maybe Natural

Specifies the requested minimum provisioned transactions (recommendations) per second.

creationDateTime :: Maybe POSIX

The date and time (in Unix format) that the campaign was created.

campaignArn :: Maybe Text

The Amazon Resource Name (ARN) of the campaign.

solutionVersionArn :: Maybe Text

The Amazon Resource Name (ARN) of a specific version of the solution.

newCampaign :: Campaign

Create a value of Campaign with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:Campaign', campaign_failureReason - If a campaign fails, the reason behind the failure.

$sel:status:Campaign', campaign_status - The status of the campaign.

A campaign can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING > DELETE IN_PROGRESS

$sel:lastUpdatedDateTime:Campaign', campaign_lastUpdatedDateTime - The date and time (in Unix format) that the campaign was last updated.

$sel:campaignConfig:Campaign', campaign_campaignConfig - The configuration details of a campaign.

$sel:latestCampaignUpdate:Campaign', campaign_latestCampaignUpdate - Undocumented member.

$sel:name:Campaign', campaign_name - The name of the campaign.

$sel:minProvisionedTPS:Campaign', campaign_minProvisionedTPS - Specifies the requested minimum provisioned transactions (recommendations) per second.

$sel:creationDateTime:Campaign', campaign_creationDateTime - The date and time (in Unix format) that the campaign was created.

$sel:campaignArn:Campaign', campaign_campaignArn - The Amazon Resource Name (ARN) of the campaign.

$sel:solutionVersionArn:Campaign', campaign_solutionVersionArn - The Amazon Resource Name (ARN) of a specific version of the solution.

campaign_failureReason :: Lens' Campaign (Maybe Text)

If a campaign fails, the reason behind the failure.

campaign_status :: Lens' Campaign (Maybe Text)

The status of the campaign.

A campaign can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING > DELETE IN_PROGRESS

campaign_lastUpdatedDateTime :: Lens' Campaign (Maybe UTCTime)

The date and time (in Unix format) that the campaign was last updated.

campaign_campaignConfig :: Lens' Campaign (Maybe CampaignConfig)

The configuration details of a campaign.

campaign_latestCampaignUpdate :: Lens' Campaign (Maybe CampaignUpdateSummary)

Undocumented member.

campaign_name :: Lens' Campaign (Maybe Text)

The name of the campaign.

campaign_minProvisionedTPS :: Lens' Campaign (Maybe Natural)

Specifies the requested minimum provisioned transactions (recommendations) per second.

campaign_creationDateTime :: Lens' Campaign (Maybe UTCTime)

The date and time (in Unix format) that the campaign was created.

campaign_campaignArn :: Lens' Campaign (Maybe Text)

The Amazon Resource Name (ARN) of the campaign.

campaign_solutionVersionArn :: Lens' Campaign (Maybe Text)

The Amazon Resource Name (ARN) of a specific version of the solution.