Amazonka.Personalize.Types.DatasetGroupSummary

Description

Documentation

data DatasetGroupSummary

Provides a summary of the properties of a dataset group. For a complete listing, call the DescribeDatasetGroup API.

See: newDatasetGroupSummary smart constructor.

Constructors

DatasetGroupSummary'

Fields

failureReason :: Maybe Text

If creating a dataset group fails, the reason behind the failure.

status :: Maybe Text

The status of the dataset group.

A dataset group can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING

lastUpdatedDateTime :: Maybe POSIX

The date and time (in Unix time) that the dataset group was last updated.

name :: Maybe Text

The name of the dataset group.

creationDateTime :: Maybe POSIX

The date and time (in Unix time) that the dataset group was created.

datasetGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset group.

newDatasetGroupSummary :: DatasetGroupSummary

Create a value of DatasetGroupSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DatasetGroupSummary', datasetGroupSummary_failureReason - If creating a dataset group fails, the reason behind the failure.

$sel:status:DatasetGroupSummary', datasetGroupSummary_status - The status of the dataset group.

A dataset group can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING

$sel:lastUpdatedDateTime:DatasetGroupSummary', datasetGroupSummary_lastUpdatedDateTime - The date and time (in Unix time) that the dataset group was last updated.

$sel:name:DatasetGroupSummary', datasetGroupSummary_name - The name of the dataset group.

$sel:creationDateTime:DatasetGroupSummary', datasetGroupSummary_creationDateTime - The date and time (in Unix time) that the dataset group was created.

$sel:datasetGroupArn:DatasetGroupSummary', datasetGroupSummary_datasetGroupArn - The Amazon Resource Name (ARN) of the dataset group.

datasetGroupSummary_failureReason :: Lens' DatasetGroupSummary (Maybe Text)

If creating a dataset group fails, the reason behind the failure.

datasetGroupSummary_status :: Lens' DatasetGroupSummary (Maybe Text)

The status of the dataset group.

A dataset group can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING

datasetGroupSummary_lastUpdatedDateTime :: Lens' DatasetGroupSummary (Maybe UTCTime)

The date and time (in Unix time) that the dataset group was last updated.

datasetGroupSummary_name :: Lens' DatasetGroupSummary (Maybe Text)

The name of the dataset group.

datasetGroupSummary_creationDateTime :: Lens' DatasetGroupSummary (Maybe UTCTime)

The date and time (in Unix time) that the dataset group was created.

datasetGroupSummary_datasetGroupArn :: Lens' DatasetGroupSummary (Maybe Text)

The Amazon Resource Name (ARN) of the dataset group.