Amazonka.Personalize.Types.Filter

Description

Documentation

data Filter

Contains information on a recommendation filter, including its ARN, status, and filter expression.

See: newFilter smart constructor.

Constructors

Filter'

Fields

failureReason :: Maybe Text

If the filter failed, the reason for its failure.

status :: Maybe Text

The status of the filter.

filterExpression :: Maybe (Sensitive Text)

Specifies the type of item interactions to filter out of recommendation results. The filter expression must follow specific format rules. For information about filter expression structure and syntax, see filter-expressions.

lastUpdatedDateTime :: Maybe POSIX

The time at which the filter was last updated.

name :: Maybe Text

The name of the filter.

filterArn :: Maybe Text

The ARN of the filter.

creationDateTime :: Maybe POSIX

The time at which the filter was created.

datasetGroupArn :: Maybe Text

The ARN of the dataset group to which the filter belongs.

newFilter :: Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:Filter', filter_failureReason - If the filter failed, the reason for its failure.

$sel:status:Filter', filter_status - The status of the filter.

$sel:filterExpression:Filter', filter_filterExpression - Specifies the type of item interactions to filter out of recommendation results. The filter expression must follow specific format rules. For information about filter expression structure and syntax, see filter-expressions.

$sel:lastUpdatedDateTime:Filter', filter_lastUpdatedDateTime - The time at which the filter was last updated.

$sel:name:Filter', filter_name - The name of the filter.

$sel:filterArn:Filter', filter_filterArn - The ARN of the filter.

$sel:creationDateTime:Filter', filter_creationDateTime - The time at which the filter was created.

$sel:datasetGroupArn:Filter', filter_datasetGroupArn - The ARN of the dataset group to which the filter belongs.

filter_failureReason :: Lens' Filter (Maybe Text)

If the filter failed, the reason for its failure.

filter_status :: Lens' Filter (Maybe Text)

The status of the filter.

filter_filterExpression :: Lens' Filter (Maybe Text)

Specifies the type of item interactions to filter out of recommendation results. The filter expression must follow specific format rules. For information about filter expression structure and syntax, see filter-expressions.

filter_lastUpdatedDateTime :: Lens' Filter (Maybe UTCTime)

The time at which the filter was last updated.

filter_name :: Lens' Filter (Maybe Text)

The name of the filter.

filter_filterArn :: Lens' Filter (Maybe Text)

The ARN of the filter.

filter_creationDateTime :: Lens' Filter (Maybe UTCTime)

The time at which the filter was created.

filter_datasetGroupArn :: Lens' Filter (Maybe Text)

The ARN of the dataset group to which the filter belongs.