Amazonka.Personalize.Types.SolutionVersion

Description

Documentation

data SolutionVersion

An object that provides information about a specific version of a Solution.

See: newSolutionVersion smart constructor.

Constructors

SolutionVersion'

Fields

failureReason :: Maybe Text

If training a solution version fails, the reason for the failure.

solutionArn :: Maybe Text

The ARN of the solution.

status :: Maybe Text

The status of the solution version.

A solution version can be in one of the following states:

CREATE PENDING

CREATE IN_PROGRESS

ACTIVE

CREATE FAILED

CREATE STOPPING

CREATE STOPPED

performAutoML :: Maybe Bool

When true, Amazon Personalize searches for the most optimal recipe according to the solution configuration. When false (the default), Amazon Personalize uses recipeArn.

tunedHPOParams :: Maybe TunedHPOParams

If hyperparameter optimization was performed, contains the hyperparameter values of the best performing model.

recipeArn :: Maybe Text

The ARN of the recipe used in the solution.

lastUpdatedDateTime :: Maybe POSIX

The date and time (in Unix time) that the solution was last updated.

eventType :: Maybe Text

The event type (for example, 'click' or 'like') that is used for training the model.

creationDateTime :: Maybe POSIX

The date and time (in Unix time) that this version of the solution was created.

datasetGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset group providing the training data.

trainingMode :: Maybe TrainingMode

The scope of training to be performed when creating the solution version. The FULL option trains the solution version based on the entirety of the input solution's training data, while the UPDATE option processes only the data that has changed in comparison to the input solution. Choose UPDATE when you want to incrementally update your solution version instead of creating an entirely new one.

The UPDATE option can only be used when you already have an active solution version created from the input solution using the FULL option and the input solution was trained with the User-Personalization recipe or the HRNN-Coldstart recipe.

trainingHours :: Maybe Double

The time used to train the model. You are billed for the time it takes to train a model. This field is visible only after Amazon Personalize successfully trains a model.

solutionConfig :: Maybe SolutionConfig

Describes the configuration properties for the solution.

performHPO :: Maybe Bool

Whether to perform hyperparameter optimization (HPO) on the chosen recipe. The default is false.

solutionVersionArn :: Maybe Text

The ARN of the solution version.

newSolutionVersion :: SolutionVersion

Create a value of SolutionVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:SolutionVersion', solutionVersion_failureReason - If training a solution version fails, the reason for the failure.

$sel:solutionArn:SolutionVersion', solutionVersion_solutionArn - The ARN of the solution.

$sel:status:SolutionVersion', solutionVersion_status - The status of the solution version.

A solution version can be in one of the following states:

CREATE PENDING

CREATE IN_PROGRESS

ACTIVE

CREATE FAILED

CREATE STOPPING

CREATE STOPPED

$sel:performAutoML:SolutionVersion', solutionVersion_performAutoML - When true, Amazon Personalize searches for the most optimal recipe according to the solution configuration. When false (the default), Amazon Personalize uses recipeArn.

$sel:tunedHPOParams:SolutionVersion', solutionVersion_tunedHPOParams - If hyperparameter optimization was performed, contains the hyperparameter values of the best performing model.

$sel:recipeArn:SolutionVersion', solutionVersion_recipeArn - The ARN of the recipe used in the solution.

$sel:lastUpdatedDateTime:SolutionVersion', solutionVersion_lastUpdatedDateTime - The date and time (in Unix time) that the solution was last updated.

$sel:eventType:SolutionVersion', solutionVersion_eventType - The event type (for example, 'click' or 'like') that is used for training the model.

$sel:creationDateTime:SolutionVersion', solutionVersion_creationDateTime - The date and time (in Unix time) that this version of the solution was created.

$sel:datasetGroupArn:SolutionVersion', solutionVersion_datasetGroupArn - The Amazon Resource Name (ARN) of the dataset group providing the training data.

$sel:trainingMode:SolutionVersion', solutionVersion_trainingMode - The scope of training to be performed when creating the solution version. The FULL option trains the solution version based on the entirety of the input solution's training data, while the UPDATE option processes only the data that has changed in comparison to the input solution. Choose UPDATE when you want to incrementally update your solution version instead of creating an entirely new one.

The UPDATE option can only be used when you already have an active solution version created from the input solution using the FULL option and the input solution was trained with the User-Personalization recipe or the HRNN-Coldstart recipe.

$sel:trainingHours:SolutionVersion', solutionVersion_trainingHours - The time used to train the model. You are billed for the time it takes to train a model. This field is visible only after Amazon Personalize successfully trains a model.

$sel:solutionConfig:SolutionVersion', solutionVersion_solutionConfig - Describes the configuration properties for the solution.

$sel:performHPO:SolutionVersion', solutionVersion_performHPO - Whether to perform hyperparameter optimization (HPO) on the chosen recipe. The default is false.

$sel:solutionVersionArn:SolutionVersion', solutionVersion_solutionVersionArn - The ARN of the solution version.

solutionVersion_failureReason :: Lens' SolutionVersion (Maybe Text)

If training a solution version fails, the reason for the failure.

solutionVersion_solutionArn :: Lens' SolutionVersion (Maybe Text)

The ARN of the solution.

solutionVersion_status :: Lens' SolutionVersion (Maybe Text)

The status of the solution version.

A solution version can be in one of the following states:

CREATE PENDING

CREATE IN_PROGRESS

ACTIVE

CREATE FAILED

CREATE STOPPING

CREATE STOPPED

solutionVersion_performAutoML :: Lens' SolutionVersion (Maybe Bool)

When true, Amazon Personalize searches for the most optimal recipe according to the solution configuration. When false (the default), Amazon Personalize uses recipeArn.

solutionVersion_tunedHPOParams :: Lens' SolutionVersion (Maybe TunedHPOParams)

If hyperparameter optimization was performed, contains the hyperparameter values of the best performing model.

solutionVersion_recipeArn :: Lens' SolutionVersion (Maybe Text)

The ARN of the recipe used in the solution.

solutionVersion_lastUpdatedDateTime :: Lens' SolutionVersion (Maybe UTCTime)

The date and time (in Unix time) that the solution was last updated.

solutionVersion_eventType :: Lens' SolutionVersion (Maybe Text)

The event type (for example, 'click' or 'like') that is used for training the model.

solutionVersion_creationDateTime :: Lens' SolutionVersion (Maybe UTCTime)

The date and time (in Unix time) that this version of the solution was created.

solutionVersion_datasetGroupArn :: Lens' SolutionVersion (Maybe Text)

The Amazon Resource Name (ARN) of the dataset group providing the training data.

solutionVersion_trainingMode :: Lens' SolutionVersion (Maybe TrainingMode)

The scope of training to be performed when creating the solution version. The FULL option trains the solution version based on the entirety of the input solution's training data, while the UPDATE option processes only the data that has changed in comparison to the input solution. Choose UPDATE when you want to incrementally update your solution version instead of creating an entirely new one.

The UPDATE option can only be used when you already have an active solution version created from the input solution using the FULL option and the input solution was trained with the User-Personalization recipe or the HRNN-Coldstart recipe.

solutionVersion_trainingHours :: Lens' SolutionVersion (Maybe Double)

The time used to train the model. You are billed for the time it takes to train a model. This field is visible only after Amazon Personalize successfully trains a model.

solutionVersion_solutionConfig :: Lens' SolutionVersion (Maybe SolutionConfig)

Describes the configuration properties for the solution.

solutionVersion_performHPO :: Lens' SolutionVersion (Maybe Bool)

Whether to perform hyperparameter optimization (HPO) on the chosen recipe. The default is false.

solutionVersion_solutionVersionArn :: Lens' SolutionVersion (Maybe Text)

The ARN of the solution version.