Amazonka.Personalize.Types.SolutionVersionSummary

Description

Documentation

data SolutionVersionSummary

Provides a summary of the properties of a solution version. For a complete listing, call the DescribeSolutionVersion API.

See: newSolutionVersionSummary smart constructor.

Constructors

SolutionVersionSummary'

Fields

failureReason :: Maybe Text

If a solution version fails, the reason behind the failure.

status :: Maybe Text

The status of the solution version.

A solution version can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

lastUpdatedDateTime :: Maybe POSIX

The date and time (in Unix time) that the solution version was last updated.

creationDateTime :: Maybe POSIX

The date and time (in Unix time) that this version of a solution was created.

solutionVersionArn :: Maybe Text

The Amazon Resource Name (ARN) of the solution version.

newSolutionVersionSummary :: SolutionVersionSummary

Create a value of SolutionVersionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:SolutionVersionSummary', solutionVersionSummary_failureReason - If a solution version fails, the reason behind the failure.

$sel:status:SolutionVersionSummary', solutionVersionSummary_status - The status of the solution version.

A solution version can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

$sel:lastUpdatedDateTime:SolutionVersionSummary', solutionVersionSummary_lastUpdatedDateTime - The date and time (in Unix time) that the solution version was last updated.

$sel:creationDateTime:SolutionVersionSummary', solutionVersionSummary_creationDateTime - The date and time (in Unix time) that this version of a solution was created.

$sel:solutionVersionArn:SolutionVersionSummary', solutionVersionSummary_solutionVersionArn - The Amazon Resource Name (ARN) of the solution version.

solutionVersionSummary_failureReason :: Lens' SolutionVersionSummary (Maybe Text)

If a solution version fails, the reason behind the failure.

solutionVersionSummary_status :: Lens' SolutionVersionSummary (Maybe Text)

The status of the solution version.

A solution version can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

solutionVersionSummary_lastUpdatedDateTime :: Lens' SolutionVersionSummary (Maybe UTCTime)

The date and time (in Unix time) that the solution version was last updated.

solutionVersionSummary_creationDateTime :: Lens' SolutionVersionSummary (Maybe UTCTime)

The date and time (in Unix time) that this version of a solution was created.

solutionVersionSummary_solutionVersionArn :: Lens' SolutionVersionSummary (Maybe Text)

The Amazon Resource Name (ARN) of the solution version.