Amazonka.PersonalizeEvents

Description

Derived from API version 2018-03-22 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Personalize can consume real-time user event data, such as stream or click data, and use it for model training either alone or combined with historical data. For more information see Recording Events.

Service Configuration

defaultService :: Service

API version 2018-03-22 of the Amazon Personalize Events SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by PersonalizeEvents.

InvalidInputException

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

Provide a valid value for the field or parameter.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Could not find the specified resource.

ResourceInUseException

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource is in use.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

PutUsers

data PutUsers

See: newPutUsers smart constructor.

Constructors

PutUsers' Text (NonEmpty User)

newPutUsers

Arguments

:: Text
$sel:datasetArn:PutUsers'

-> NonEmpty User
$sel:users:PutUsers'

-> PutUsers

Create a value of PutUsers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetArn:PutUsers', putUsers_datasetArn - The Amazon Resource Name (ARN) of the Users dataset you are adding the user or users to.

$sel:users:PutUsers', putUsers_users - A list of user data.

data PutUsersResponse

See: newPutUsersResponse smart constructor.

Constructors

PutUsersResponse'

newPutUsersResponse :: PutUsersResponse

Create a value of PutUsersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

PutItems

data PutItems

See: newPutItems smart constructor.

Constructors

PutItems' Text (NonEmpty Item)

newPutItems

Arguments

:: Text
$sel:datasetArn:PutItems'

-> NonEmpty Item
$sel:items:PutItems'

-> PutItems

Create a value of PutItems with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetArn:PutItems', putItems_datasetArn - The Amazon Resource Name (ARN) of the Items dataset you are adding the item or items to.

$sel:items:PutItems', putItems_items - A list of item data.

data PutItemsResponse

See: newPutItemsResponse smart constructor.

Constructors

PutItemsResponse'

newPutItemsResponse :: PutItemsResponse

Create a value of PutItemsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

PutEvents

data PutEvents

See: newPutEvents smart constructor.

Constructors

PutEvents' (Maybe Text) Text Text (NonEmpty Event)

newPutEvents

Arguments

:: Text
$sel:trackingId:PutEvents'

-> Text
$sel:sessionId:PutEvents'

-> NonEmpty Event
$sel:eventList:PutEvents'

-> PutEvents

Create a value of PutEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userId:PutEvents', putEvents_userId - The user associated with the event.

$sel:trackingId:PutEvents', putEvents_trackingId - The tracking ID for the event. The ID is generated by a call to the CreateEventTracker API.

$sel:sessionId:PutEvents', putEvents_sessionId - The session ID associated with the user's visit. Your application generates the sessionId when a user first visits your website or uses your application. Amazon Personalize uses the sessionId to associate events with the user before they log in. For more information, see Recording Events.

$sel:eventList:PutEvents', putEvents_eventList - A list of event data from the session.

data PutEventsResponse

See: newPutEventsResponse smart constructor.

Constructors

PutEventsResponse'

newPutEventsResponse :: PutEventsResponse

Create a value of PutEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Types

Event

data Event

Represents user interaction event information sent using the PutEvents API.

See: newEvent smart constructor.

Constructors

Event' (Maybe Text) (Maybe Double) (Maybe Text) (Maybe (NonEmpty Text)) (Maybe Text) (Maybe Text) Text POSIX

newEvent

Arguments

:: Text
$sel:eventType:Event'

-> UTCTime
$sel:sentAt:Event'

-> Event

Create a value of Event with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommendationId:Event', event_recommendationId - The ID of the recommendation.

$sel:eventValue:Event', event_eventValue - The event value that corresponds to the EVENT_VALUE field of the Interactions schema.

$sel:itemId:Event', event_itemId - The item ID key that corresponds to the ITEM_ID field of the Interactions schema.

$sel:impression:Event', event_impression - A list of item IDs that represents the sequence of items you have shown the user. For example, ["itemId1", "itemId2", "itemId3"].

$sel:eventId:Event', event_eventId - An ID associated with the event. If an event ID is not provided, Amazon Personalize generates a unique ID for the event. An event ID is not used as an input to the model. Amazon Personalize uses the event ID to distinquish unique events. Any subsequent events after the first with the same event ID are not used in model training.

$sel:properties:Event', event_properties - A string map of event-specific data that you might choose to record. For example, if a user rates a movie on your site, other than movie ID (itemId) and rating (eventValue) , you might also send the number of movie ratings made by the user.

Each item in the map consists of a key-value pair. For example,

{"numberOfRatings": "12"}

The keys use camel case names that match the fields in the Interactions schema. In the above example, the numberOfRatings would match the 'NUMBER_OF_RATINGS' field defined in the Interactions schema.

$sel:eventType:Event', event_eventType - The type of event, such as click or download. This property corresponds to the EVENT_TYPE field of your Interactions schema and depends on the types of events you are tracking.

$sel:sentAt:Event', event_sentAt - The timestamp (in Unix time) on the client side when the event occurred.

Item

data Item

Represents item metadata added to an Items dataset using the PutItems API. For more information see Importing Items Incrementally.

See: newItem smart constructor.

Constructors

Item' (Maybe Text) Text

newItem

Arguments

:: Text
$sel:itemId:Item'

-> Item

Create a value of Item with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:properties:Item', item_properties - A string map of item-specific metadata. Each element in the map consists of a key-value pair. For example, {"numberOfRatings": "12"}.

The keys use camel case names that match the fields in the schema for the Items dataset. In the previous example, the numberOfRatings matches the 'NUMBER_OF_RATINGS' field defined in the Items schema. For categorical string data, to include multiple categories for a single item, separate each category with a pipe separator (|). For example, \"Horror|Action\".

$sel:itemId:Item', item_itemId - The ID associated with the item.

User

data User

Represents user metadata added to a Users dataset using the PutUsers API. For more information see Importing Users Incrementally.

See: newUser smart constructor.

Constructors

User' (Maybe Text) Text

newUser

Arguments

:: Text
$sel:userId:User'

-> User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:properties:User', user_properties - A string map of user-specific metadata. Each element in the map consists of a key-value pair. For example, {"numberOfVideosWatched": "45"}.

The keys use camel case names that match the fields in the schema for the Users dataset. In the previous example, the numberOfVideosWatched matches the 'NUMBER_OF_VIDEOS_WATCHED' field defined in the Users schema. For categorical string data, to include multiple categories for a single user, separate each category with a pipe separator (|). For example, \"Member|Frequent shopper\".

$sel:userId:User', user_userId - The ID associated with the user.