Amazonka.PersonalizeEvents.PutUsers

Description

Adds one or more users to a Users dataset. For more information see Importing Users Incrementally.

Creating a Request

data PutUsers

See: newPutUsers smart constructor.

Constructors

PutUsers'

Fields

datasetArn :: Text

The Amazon Resource Name (ARN) of the Users dataset you are adding the user or users to.

users :: NonEmpty User

A list of user data.

newPutUsers

Arguments

:: Text
$sel:datasetArn:PutUsers'

-> NonEmpty User
$sel:users:PutUsers'

-> PutUsers

Create a value of PutUsers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetArn:PutUsers', putUsers_datasetArn - The Amazon Resource Name (ARN) of the Users dataset you are adding the user or users to.

$sel:users:PutUsers', putUsers_users - A list of user data.

Request Lenses

putUsers_datasetArn :: Lens' PutUsers Text

The Amazon Resource Name (ARN) of the Users dataset you are adding the user or users to.

putUsers_users :: Lens' PutUsers (NonEmpty User)

A list of user data.

Destructuring the Response

data PutUsersResponse

See: newPutUsersResponse smart constructor.

Constructors

PutUsersResponse'

newPutUsersResponse :: PutUsersResponse

Create a value of PutUsersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.