Amazonka.PersonalizeEvents.Types

Description

Service Configuration

defaultService :: Service

API version 2018-03-22 of the Amazon Personalize Events SDK configuration.

Errors

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

Provide a valid value for the field or parameter.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Could not find the specified resource.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource is in use.

Event

data Event

Represents user interaction event information sent using the PutEvents API.

See: newEvent smart constructor.

Constructors

Event'

Fields

recommendationId :: Maybe Text

The ID of the recommendation.

eventValue :: Maybe Double

The event value that corresponds to the EVENT_VALUE field of the Interactions schema.

itemId :: Maybe Text

The item ID key that corresponds to the ITEM_ID field of the Interactions schema.

impression :: Maybe (NonEmpty Text)

A list of item IDs that represents the sequence of items you have shown the user. For example, ["itemId1", "itemId2", "itemId3"].

eventId :: Maybe Text

An ID associated with the event. If an event ID is not provided, Amazon Personalize generates a unique ID for the event. An event ID is not used as an input to the model. Amazon Personalize uses the event ID to distinquish unique events. Any subsequent events after the first with the same event ID are not used in model training.

properties :: Maybe Text

A string map of event-specific data that you might choose to record. For example, if a user rates a movie on your site, other than movie ID (itemId) and rating (eventValue) , you might also send the number of movie ratings made by the user.

Each item in the map consists of a key-value pair. For example,

{"numberOfRatings": "12"}

The keys use camel case names that match the fields in the Interactions schema. In the above example, the numberOfRatings would match the 'NUMBER_OF_RATINGS' field defined in the Interactions schema.

eventType :: Text

The type of event, such as click or download. This property corresponds to the EVENT_TYPE field of your Interactions schema and depends on the types of events you are tracking.

sentAt :: POSIX

The timestamp (in Unix time) on the client side when the event occurred.

newEvent

Arguments

:: Text
$sel:eventType:Event'

-> UTCTime
$sel:sentAt:Event'

-> Event

Create a value of Event with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommendationId:Event', event_recommendationId - The ID of the recommendation.

$sel:eventValue:Event', event_eventValue - The event value that corresponds to the EVENT_VALUE field of the Interactions schema.

$sel:itemId:Event', event_itemId - The item ID key that corresponds to the ITEM_ID field of the Interactions schema.

$sel:impression:Event', event_impression - A list of item IDs that represents the sequence of items you have shown the user. For example, ["itemId1", "itemId2", "itemId3"].

$sel:eventId:Event', event_eventId - An ID associated with the event. If an event ID is not provided, Amazon Personalize generates a unique ID for the event. An event ID is not used as an input to the model. Amazon Personalize uses the event ID to distinquish unique events. Any subsequent events after the first with the same event ID are not used in model training.

$sel:properties:Event', event_properties - A string map of event-specific data that you might choose to record. For example, if a user rates a movie on your site, other than movie ID (itemId) and rating (eventValue) , you might also send the number of movie ratings made by the user.

Each item in the map consists of a key-value pair. For example,

{"numberOfRatings": "12"}

The keys use camel case names that match the fields in the Interactions schema. In the above example, the numberOfRatings would match the 'NUMBER_OF_RATINGS' field defined in the Interactions schema.

$sel:eventType:Event', event_eventType - The type of event, such as click or download. This property corresponds to the EVENT_TYPE field of your Interactions schema and depends on the types of events you are tracking.

$sel:sentAt:Event', event_sentAt - The timestamp (in Unix time) on the client side when the event occurred.

event_recommendationId :: Lens' Event (Maybe Text)

The ID of the recommendation.

event_eventValue :: Lens' Event (Maybe Double)

The event value that corresponds to the EVENT_VALUE field of the Interactions schema.

event_itemId :: Lens' Event (Maybe Text)

The item ID key that corresponds to the ITEM_ID field of the Interactions schema.

event_impression :: Lens' Event (Maybe (NonEmpty Text))

A list of item IDs that represents the sequence of items you have shown the user. For example, ["itemId1", "itemId2", "itemId3"].

event_eventId :: Lens' Event (Maybe Text)

An ID associated with the event. If an event ID is not provided, Amazon Personalize generates a unique ID for the event. An event ID is not used as an input to the model. Amazon Personalize uses the event ID to distinquish unique events. Any subsequent events after the first with the same event ID are not used in model training.

event_properties :: Lens' Event (Maybe Text)

A string map of event-specific data that you might choose to record. For example, if a user rates a movie on your site, other than movie ID (itemId) and rating (eventValue) , you might also send the number of movie ratings made by the user.

Each item in the map consists of a key-value pair. For example,

{"numberOfRatings": "12"}

The keys use camel case names that match the fields in the Interactions schema. In the above example, the numberOfRatings would match the 'NUMBER_OF_RATINGS' field defined in the Interactions schema.

event_eventType :: Lens' Event Text

The type of event, such as click or download. This property corresponds to the EVENT_TYPE field of your Interactions schema and depends on the types of events you are tracking.

event_sentAt :: Lens' Event UTCTime

The timestamp (in Unix time) on the client side when the event occurred.

Item

data Item

Represents item metadata added to an Items dataset using the PutItems API. For more information see Importing Items Incrementally.

See: newItem smart constructor.

Constructors

Item'

Fields

properties :: Maybe Text

A string map of item-specific metadata. Each element in the map consists of a key-value pair. For example, {"numberOfRatings": "12"}.

The keys use camel case names that match the fields in the schema for the Items dataset. In the previous example, the numberOfRatings matches the 'NUMBER_OF_RATINGS' field defined in the Items schema. For categorical string data, to include multiple categories for a single item, separate each category with a pipe separator (|). For example, \"Horror|Action\".

itemId :: Text

The ID associated with the item.

newItem

Arguments

:: Text
$sel:itemId:Item'

-> Item

Create a value of Item with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:properties:Item', item_properties - A string map of item-specific metadata. Each element in the map consists of a key-value pair. For example, {"numberOfRatings": "12"}.

The keys use camel case names that match the fields in the schema for the Items dataset. In the previous example, the numberOfRatings matches the 'NUMBER_OF_RATINGS' field defined in the Items schema. For categorical string data, to include multiple categories for a single item, separate each category with a pipe separator (|). For example, \"Horror|Action\".

$sel:itemId:Item', item_itemId - The ID associated with the item.

item_properties :: Lens' Item (Maybe Text)

A string map of item-specific metadata. Each element in the map consists of a key-value pair. For example, {"numberOfRatings": "12"}.

The keys use camel case names that match the fields in the schema for the Items dataset. In the previous example, the numberOfRatings matches the 'NUMBER_OF_RATINGS' field defined in the Items schema. For categorical string data, to include multiple categories for a single item, separate each category with a pipe separator (|). For example, \"Horror|Action\".

item_itemId :: Lens' Item Text

The ID associated with the item.

User

data User

Represents user metadata added to a Users dataset using the PutUsers API. For more information see Importing Users Incrementally.

See: newUser smart constructor.

Constructors

User'

Fields

properties :: Maybe Text

A string map of user-specific metadata. Each element in the map consists of a key-value pair. For example, {"numberOfVideosWatched": "45"}.

The keys use camel case names that match the fields in the schema for the Users dataset. In the previous example, the numberOfVideosWatched matches the 'NUMBER_OF_VIDEOS_WATCHED' field defined in the Users schema. For categorical string data, to include multiple categories for a single user, separate each category with a pipe separator (|). For example, \"Member|Frequent shopper\".

userId :: Text

The ID associated with the user.

newUser

Arguments

:: Text
$sel:userId:User'

-> User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:properties:User', user_properties - A string map of user-specific metadata. Each element in the map consists of a key-value pair. For example, {"numberOfVideosWatched": "45"}.

The keys use camel case names that match the fields in the schema for the Users dataset. In the previous example, the numberOfVideosWatched matches the 'NUMBER_OF_VIDEOS_WATCHED' field defined in the Users schema. For categorical string data, to include multiple categories for a single user, separate each category with a pipe separator (|). For example, \"Member|Frequent shopper\".

$sel:userId:User', user_userId - The ID associated with the user.

user_properties :: Lens' User (Maybe Text)

A string map of user-specific metadata. Each element in the map consists of a key-value pair. For example, {"numberOfVideosWatched": "45"}.

The keys use camel case names that match the fields in the schema for the Users dataset. In the previous example, the numberOfVideosWatched matches the 'NUMBER_OF_VIDEOS_WATCHED' field defined in the Users schema. For categorical string data, to include multiple categories for a single user, separate each category with a pipe separator (|). For example, \"Member|Frequent shopper\".

user_userId :: Lens' User Text

The ID associated with the user.