Amazonka.Pi.DescribeDimensionKeys

Description

For a specific time period, retrieve the top N dimension keys for a metric.

Each response element returns a maximum of 500 bytes. For larger elements, such as SQL statements, only the first 500 bytes are returned.

Creating a Request

data DescribeDimensionKeys

See: newDescribeDimensionKeys smart constructor.

Constructors

DescribeDimensionKeys'

Fields

periodInSeconds :: Maybe Int

The granularity, in seconds, of the data points returned from Performance Insights. A period can be as short as one second, or as long as one day (86400 seconds). Valid values are:

1 (one second)

60 (one minute)

300 (five minutes)

3600 (one hour)

86400 (twenty-four hours)

If you don't specify PeriodInSeconds, then Performance Insights chooses a value for you, with a goal of returning roughly 100-200 data points in the response.

nextToken :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

filter' :: Maybe (HashMap Text Text)

One or more filters to apply in the request. Restrictions:

Any number of filters by the same dimension, as specified in the GroupBy or Partition parameters.

A single filter for any other dimension in this dimension group.

maxResults :: Maybe Natural

The maximum number of items to return in the response. If more items exist than the specified MaxRecords value, a pagination token is included in the response so that the remaining results can be retrieved.

partitionBy :: Maybe DimensionGroup

For each dimension specified in GroupBy, specify a secondary dimension to further subdivide the partition keys in the response.

serviceType :: ServiceType

The AWS service for which Performance Insights will return metrics. The only valid value for ServiceType is RDS.

identifier :: Text

An immutable, AWS Region-unique identifier for a data source. Performance Insights gathers metrics from this data source.

To use an Amazon RDS instance as a data source, you specify its DbiResourceId value. For example, specify db-FAIHNTYBKTGAUSUZQYPDS2GW4A

startTime :: POSIX

The date and time specifying the beginning of the requested time series data. You must specify a StartTime within the past 7 days. The value specified is inclusive, which means that data points equal to or greater than StartTime are returned.

The value for StartTime must be earlier than the value for EndTime.

endTime :: POSIX

The date and time specifying the end of the requested time series data. The value specified is exclusive, which means that data points less than (but not equal to) EndTime are returned.

The value for EndTime must be later than the value for StartTime.

metric :: Text

The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

groupBy :: DimensionGroup

A specification for how to aggregate the data points from a query result. You must specify a valid dimension group. Performance Insights returns all dimensions within this group, unless you provide the names of specific dimensions within this group. You can also request that Performance Insights return a limited number of values for a dimension.

newDescribeDimensionKeys

Arguments

:: ServiceType
$sel:serviceType:DescribeDimensionKeys'

-> Text
$sel:identifier:DescribeDimensionKeys'

-> UTCTime
$sel:startTime:DescribeDimensionKeys'

-> UTCTime
$sel:endTime:DescribeDimensionKeys'

-> Text
$sel:metric:DescribeDimensionKeys'

-> DimensionGroup
$sel:groupBy:DescribeDimensionKeys'

-> DescribeDimensionKeys

Create a value of DescribeDimensionKeys with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:periodInSeconds:DescribeDimensionKeys', describeDimensionKeys_periodInSeconds - The granularity, in seconds, of the data points returned from Performance Insights. A period can be as short as one second, or as long as one day (86400 seconds). Valid values are:

1 (one second)

60 (one minute)

300 (five minutes)

3600 (one hour)

86400 (twenty-four hours)

If you don't specify PeriodInSeconds, then Performance Insights chooses a value for you, with a goal of returning roughly 100-200 data points in the response.

$sel:nextToken:DescribeDimensionKeys', describeDimensionKeys_nextToken - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

$sel:filter':DescribeDimensionKeys', describeDimensionKeys_filter - One or more filters to apply in the request. Restrictions:

Any number of filters by the same dimension, as specified in the GroupBy or Partition parameters.

A single filter for any other dimension in this dimension group.

$sel:maxResults:DescribeDimensionKeys', describeDimensionKeys_maxResults - The maximum number of items to return in the response. If more items exist than the specified MaxRecords value, a pagination token is included in the response so that the remaining results can be retrieved.

$sel:partitionBy:DescribeDimensionKeys', describeDimensionKeys_partitionBy - For each dimension specified in GroupBy, specify a secondary dimension to further subdivide the partition keys in the response.

$sel:serviceType:DescribeDimensionKeys', describeDimensionKeys_serviceType - The AWS service for which Performance Insights will return metrics. The only valid value for ServiceType is RDS.

$sel:identifier:DescribeDimensionKeys', describeDimensionKeys_identifier - An immutable, AWS Region-unique identifier for a data source. Performance Insights gathers metrics from this data source.

To use an Amazon RDS instance as a data source, you specify its DbiResourceId value. For example, specify db-FAIHNTYBKTGAUSUZQYPDS2GW4A

$sel:startTime:DescribeDimensionKeys', describeDimensionKeys_startTime - The date and time specifying the beginning of the requested time series data. You must specify a StartTime within the past 7 days. The value specified is inclusive, which means that data points equal to or greater than StartTime are returned.

The value for StartTime must be earlier than the value for EndTime.

$sel:endTime:DescribeDimensionKeys', describeDimensionKeys_endTime - The date and time specifying the end of the requested time series data. The value specified is exclusive, which means that data points less than (but not equal to) EndTime are returned.

The value for EndTime must be later than the value for StartTime.

$sel:metric:DescribeDimensionKeys', describeDimensionKeys_metric - The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

$sel:groupBy:DescribeDimensionKeys', describeDimensionKeys_groupBy - A specification for how to aggregate the data points from a query result. You must specify a valid dimension group. Performance Insights returns all dimensions within this group, unless you provide the names of specific dimensions within this group. You can also request that Performance Insights return a limited number of values for a dimension.

Request Lenses

describeDimensionKeys_periodInSeconds :: Lens' DescribeDimensionKeys (Maybe Int)

The granularity, in seconds, of the data points returned from Performance Insights. A period can be as short as one second, or as long as one day (86400 seconds). Valid values are:

1 (one second)

60 (one minute)

300 (five minutes)

3600 (one hour)

86400 (twenty-four hours)

If you don't specify PeriodInSeconds, then Performance Insights chooses a value for you, with a goal of returning roughly 100-200 data points in the response.

describeDimensionKeys_nextToken :: Lens' DescribeDimensionKeys (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

describeDimensionKeys_filter :: Lens' DescribeDimensionKeys (Maybe (HashMap Text Text))

One or more filters to apply in the request. Restrictions:

Any number of filters by the same dimension, as specified in the GroupBy or Partition parameters.

A single filter for any other dimension in this dimension group.

describeDimensionKeys_maxResults :: Lens' DescribeDimensionKeys (Maybe Natural)

The maximum number of items to return in the response. If more items exist than the specified MaxRecords value, a pagination token is included in the response so that the remaining results can be retrieved.

describeDimensionKeys_partitionBy :: Lens' DescribeDimensionKeys (Maybe DimensionGroup)

For each dimension specified in GroupBy, specify a secondary dimension to further subdivide the partition keys in the response.

describeDimensionKeys_serviceType :: Lens' DescribeDimensionKeys ServiceType

The AWS service for which Performance Insights will return metrics. The only valid value for ServiceType is RDS.

describeDimensionKeys_identifier :: Lens' DescribeDimensionKeys Text

An immutable, AWS Region-unique identifier for a data source. Performance Insights gathers metrics from this data source.

To use an Amazon RDS instance as a data source, you specify its DbiResourceId value. For example, specify db-FAIHNTYBKTGAUSUZQYPDS2GW4A

describeDimensionKeys_startTime :: Lens' DescribeDimensionKeys UTCTime

The date and time specifying the beginning of the requested time series data. You must specify a StartTime within the past 7 days. The value specified is inclusive, which means that data points equal to or greater than StartTime are returned.

The value for StartTime must be earlier than the value for EndTime.

describeDimensionKeys_endTime :: Lens' DescribeDimensionKeys UTCTime

The date and time specifying the end of the requested time series data. The value specified is exclusive, which means that data points less than (but not equal to) EndTime are returned.

The value for EndTime must be later than the value for StartTime.

describeDimensionKeys_metric :: Lens' DescribeDimensionKeys Text

The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

describeDimensionKeys_groupBy :: Lens' DescribeDimensionKeys DimensionGroup

A specification for how to aggregate the data points from a query result. You must specify a valid dimension group. Performance Insights returns all dimensions within this group, unless you provide the names of specific dimensions within this group. You can also request that Performance Insights return a limited number of values for a dimension.

Destructuring the Response

data DescribeDimensionKeysResponse

See: newDescribeDimensionKeysResponse smart constructor.

Constructors

DescribeDimensionKeysResponse'

Fields

alignedEndTime :: Maybe POSIX

The end time for the returned dimension keys, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedEndTime will be greater than or equal to the value of the user-specified Endtime.

alignedStartTime :: Maybe POSIX

The start time for the returned dimension keys, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedStartTime will be less than or equal to the value of the user-specified StartTime.

keys :: Maybe [DimensionKeyDescription]

The dimension keys that were requested.

nextToken :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

partitionKeys :: Maybe [ResponsePartitionKey]

If PartitionBy was present in the request, PartitionKeys contains the breakdown of dimension keys by the specified partitions.

httpStatus :: Int

The response's http status code.

newDescribeDimensionKeysResponse

Arguments

:: Int
$sel:httpStatus:DescribeDimensionKeysResponse'

-> DescribeDimensionKeysResponse

Create a value of DescribeDimensionKeysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alignedEndTime:DescribeDimensionKeysResponse', describeDimensionKeysResponse_alignedEndTime - The end time for the returned dimension keys, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedEndTime will be greater than or equal to the value of the user-specified Endtime.

$sel:alignedStartTime:DescribeDimensionKeysResponse', describeDimensionKeysResponse_alignedStartTime - The start time for the returned dimension keys, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedStartTime will be less than or equal to the value of the user-specified StartTime.

$sel:keys:DescribeDimensionKeysResponse', describeDimensionKeysResponse_keys - The dimension keys that were requested.

$sel:nextToken:DescribeDimensionKeys', describeDimensionKeysResponse_nextToken - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

$sel:partitionKeys:DescribeDimensionKeysResponse', describeDimensionKeysResponse_partitionKeys - If PartitionBy was present in the request, PartitionKeys contains the breakdown of dimension keys by the specified partitions.

$sel:httpStatus:DescribeDimensionKeysResponse', describeDimensionKeysResponse_httpStatus - The response's http status code.

Response Lenses

describeDimensionKeysResponse_alignedEndTime :: Lens' DescribeDimensionKeysResponse (Maybe UTCTime)

The end time for the returned dimension keys, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedEndTime will be greater than or equal to the value of the user-specified Endtime.

describeDimensionKeysResponse_alignedStartTime :: Lens' DescribeDimensionKeysResponse (Maybe UTCTime)

The start time for the returned dimension keys, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedStartTime will be less than or equal to the value of the user-specified StartTime.

describeDimensionKeysResponse_keys :: Lens' DescribeDimensionKeysResponse (Maybe [DimensionKeyDescription])

The dimension keys that were requested.

describeDimensionKeysResponse_nextToken :: Lens' DescribeDimensionKeysResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

describeDimensionKeysResponse_partitionKeys :: Lens' DescribeDimensionKeysResponse (Maybe [ResponsePartitionKey])

If PartitionBy was present in the request, PartitionKeys contains the breakdown of dimension keys by the specified partitions.

describeDimensionKeysResponse_httpStatus :: Lens' DescribeDimensionKeysResponse Int

The response's http status code.