Amazonka.Pi.Lens

Description

Operations

GetDimensionKeyDetails

getDimensionKeyDetails_requestedDimensions :: Lens' GetDimensionKeyDetails (Maybe (NonEmpty Text))

A list of dimensions to retrieve the detail data for within the given dimension group. For the dimension group db.sql, specify either the full dimension name db.sql.statement or the short dimension name statement. If you don't specify this parameter, Performance Insights returns all dimension data within the specified dimension group.

getDimensionKeyDetails_serviceType :: Lens' GetDimensionKeyDetails ServiceType

The AWS service for which Performance Insights returns data. The only valid value is RDS.

getDimensionKeyDetails_identifier :: Lens' GetDimensionKeyDetails Text

The ID for a data source from which to gather dimension data. This ID must be immutable and unique within an AWS Region. When a DB instance is the data source, specify its DbiResourceId value. For example, specify db-ABCDEFGHIJKLMNOPQRSTU1VW2X.

getDimensionKeyDetails_group :: Lens' GetDimensionKeyDetails Text

The name of the dimension group. The only valid value is db.sql. Performance Insights searches the specified group for the dimension group ID.

getDimensionKeyDetails_groupIdentifier :: Lens' GetDimensionKeyDetails Text

The ID of the dimension group from which to retrieve dimension details. For dimension group db.sql, the group ID is db.sql.id.

getDimensionKeyDetailsResponse_dimensions :: Lens' GetDimensionKeyDetailsResponse (Maybe [DimensionKeyDetail])

The details for the requested dimensions.

getDimensionKeyDetailsResponse_httpStatus :: Lens' GetDimensionKeyDetailsResponse Int

The response's http status code.

GetResourceMetrics

getResourceMetrics_periodInSeconds :: Lens' GetResourceMetrics (Maybe Int)

The granularity, in seconds, of the data points returned from Performance Insights. A period can be as short as one second, or as long as one day (86400 seconds). Valid values are:

1 (one second)

60 (one minute)

300 (five minutes)

3600 (one hour)

86400 (twenty-four hours)

If you don't specify PeriodInSeconds, then Performance Insights will choose a value for you, with a goal of returning roughly 100-200 data points in the response.

getResourceMetrics_nextToken :: Lens' GetResourceMetrics (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

getResourceMetrics_maxResults :: Lens' GetResourceMetrics (Maybe Natural)

The maximum number of items to return in the response. If more items exist than the specified MaxRecords value, a pagination token is included in the response so that the remaining results can be retrieved.

getResourceMetrics_serviceType :: Lens' GetResourceMetrics ServiceType

The AWS service for which Performance Insights returns metrics. The only valid value for ServiceType is RDS.

getResourceMetrics_identifier :: Lens' GetResourceMetrics Text

An immutable, AWS Region-unique identifier for a data source. Performance Insights gathers metrics from this data source.

To use a DB instance as a data source, specify its DbiResourceId value. For example, specify db-FAIHNTYBKTGAUSUZQYPDS2GW4A.

getResourceMetrics_metricQueries :: Lens' GetResourceMetrics (NonEmpty MetricQuery)

An array of one or more queries to perform. Each query must specify a Performance Insights metric, and can optionally specify aggregation and filtering criteria.

getResourceMetrics_startTime :: Lens' GetResourceMetrics UTCTime

The date and time specifying the beginning of the requested time series data. You can't specify a StartTime that's earlier than 7 days ago. The value specified is inclusive - data points equal to or greater than StartTime will be returned.

The value for StartTime must be earlier than the value for EndTime.

getResourceMetrics_endTime :: Lens' GetResourceMetrics UTCTime

The date and time specifying the end of the requested time series data. The value specified is exclusive - data points less than (but not equal to) EndTime will be returned.

The value for EndTime must be later than the value for StartTime.

getResourceMetricsResponse_identifier :: Lens' GetResourceMetricsResponse (Maybe Text)

An immutable, AWS Region-unique identifier for a data source. Performance Insights gathers metrics from this data source.

To use a DB instance as a data source, you specify its DbiResourceId value - for example: db-FAIHNTYBKTGAUSUZQYPDS2GW4A

getResourceMetricsResponse_metricList :: Lens' GetResourceMetricsResponse (Maybe [MetricKeyDataPoints])

An array of metric results,, where each array element contains all of the data points for a particular dimension.

getResourceMetricsResponse_alignedEndTime :: Lens' GetResourceMetricsResponse (Maybe UTCTime)

The end time for the returned metrics, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedEndTime will be greater than or equal to the value of the user-specified Endtime.

getResourceMetricsResponse_alignedStartTime :: Lens' GetResourceMetricsResponse (Maybe UTCTime)

The start time for the returned metrics, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedStartTime will be less than or equal to the value of the user-specified StartTime.

getResourceMetricsResponse_nextToken :: Lens' GetResourceMetricsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

getResourceMetricsResponse_httpStatus :: Lens' GetResourceMetricsResponse Int

The response's http status code.

DescribeDimensionKeys

describeDimensionKeys_periodInSeconds :: Lens' DescribeDimensionKeys (Maybe Int)

The granularity, in seconds, of the data points returned from Performance Insights. A period can be as short as one second, or as long as one day (86400 seconds). Valid values are:

1 (one second)

60 (one minute)

300 (five minutes)

3600 (one hour)

86400 (twenty-four hours)

If you don't specify PeriodInSeconds, then Performance Insights chooses a value for you, with a goal of returning roughly 100-200 data points in the response.

describeDimensionKeys_nextToken :: Lens' DescribeDimensionKeys (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

describeDimensionKeys_filter :: Lens' DescribeDimensionKeys (Maybe (HashMap Text Text))

One or more filters to apply in the request. Restrictions:

Any number of filters by the same dimension, as specified in the GroupBy or Partition parameters.

A single filter for any other dimension in this dimension group.

describeDimensionKeys_maxResults :: Lens' DescribeDimensionKeys (Maybe Natural)

The maximum number of items to return in the response. If more items exist than the specified MaxRecords value, a pagination token is included in the response so that the remaining results can be retrieved.

describeDimensionKeys_partitionBy :: Lens' DescribeDimensionKeys (Maybe DimensionGroup)

For each dimension specified in GroupBy, specify a secondary dimension to further subdivide the partition keys in the response.

describeDimensionKeys_serviceType :: Lens' DescribeDimensionKeys ServiceType

The AWS service for which Performance Insights will return metrics. The only valid value for ServiceType is RDS.

describeDimensionKeys_identifier :: Lens' DescribeDimensionKeys Text

An immutable, AWS Region-unique identifier for a data source. Performance Insights gathers metrics from this data source.

To use an Amazon RDS instance as a data source, you specify its DbiResourceId value. For example, specify db-FAIHNTYBKTGAUSUZQYPDS2GW4A

describeDimensionKeys_startTime :: Lens' DescribeDimensionKeys UTCTime

The date and time specifying the beginning of the requested time series data. You must specify a StartTime within the past 7 days. The value specified is inclusive, which means that data points equal to or greater than StartTime are returned.

The value for StartTime must be earlier than the value for EndTime.

describeDimensionKeys_endTime :: Lens' DescribeDimensionKeys UTCTime

The date and time specifying the end of the requested time series data. The value specified is exclusive, which means that data points less than (but not equal to) EndTime are returned.

The value for EndTime must be later than the value for StartTime.

describeDimensionKeys_metric :: Lens' DescribeDimensionKeys Text

The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

describeDimensionKeys_groupBy :: Lens' DescribeDimensionKeys DimensionGroup

A specification for how to aggregate the data points from a query result. You must specify a valid dimension group. Performance Insights returns all dimensions within this group, unless you provide the names of specific dimensions within this group. You can also request that Performance Insights return a limited number of values for a dimension.

describeDimensionKeysResponse_alignedEndTime :: Lens' DescribeDimensionKeysResponse (Maybe UTCTime)

The end time for the returned dimension keys, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedEndTime will be greater than or equal to the value of the user-specified Endtime.

describeDimensionKeysResponse_alignedStartTime :: Lens' DescribeDimensionKeysResponse (Maybe UTCTime)

The start time for the returned dimension keys, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedStartTime will be less than or equal to the value of the user-specified StartTime.

describeDimensionKeysResponse_keys :: Lens' DescribeDimensionKeysResponse (Maybe [DimensionKeyDescription])

The dimension keys that were requested.

describeDimensionKeysResponse_nextToken :: Lens' DescribeDimensionKeysResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

describeDimensionKeysResponse_partitionKeys :: Lens' DescribeDimensionKeysResponse (Maybe [ResponsePartitionKey])

If PartitionBy was present in the request, PartitionKeys contains the breakdown of dimension keys by the specified partitions.

describeDimensionKeysResponse_httpStatus :: Lens' DescribeDimensionKeysResponse Int

The response's http status code.

Types

DataPoint

dataPoint_timestamp :: Lens' DataPoint UTCTime

The time, in epoch format, associated with a particular Value.

dataPoint_value :: Lens' DataPoint Double

The actual value associated with a particular Timestamp.

DimensionGroup

dimensionGroup_limit :: Lens' DimensionGroup (Maybe Natural)

The maximum number of items to fetch for this dimension group.

dimensionGroup_dimensions :: Lens' DimensionGroup (Maybe (NonEmpty Text))

A list of specific dimensions from a dimension group. If this parameter is not present, then it signifies that all of the dimensions in the group were requested, or are present in the response.

Valid values for elements in the Dimensions array are:

db.application.name - The name of the application that is connected to the database (only Aurora PostgreSQL and RDS PostgreSQL)

db.host.id - The host ID of the connected client (all engines)

db.host.name - The host name of the connected client (all engines)

db.name - The name of the database to which the client is connected (only Aurora PostgreSQL, RDS PostgreSQL, Aurora MySQL, RDS MySQL, and MariaDB)

db.session_type.name - The type of the current session (only Aurora PostgreSQL and RDS PostgreSQL)

db.sql.id - The SQL ID generated by Performance Insights (all engines)

db.sql.db_id - The SQL ID generated by the database (all engines)

db.sql.statement - The SQL text that is being executed (all engines)

db.sql.tokenized_id

db.sql_tokenized.id - The SQL digest ID generated by Performance Insights (all engines)

db.sql_tokenized.db_id - SQL digest ID generated by the database (all engines)

db.sql_tokenized.statement - The SQL digest text (all engines)

db.user.id - The ID of the user logged in to the database (all engines)

db.user.name - The name of the user logged in to the database (all engines)

db.wait_event.name - The event for which the backend is waiting (all engines)

db.wait_event.type - The type of event for which the backend is waiting (all engines)

db.wait_event_type.name - The name of the event type for which the backend is waiting (all engines)

dimensionGroup_group :: Lens' DimensionGroup Text

The name of the dimension group. Valid values are:

db - The name of the database to which the client is connected (only Aurora PostgreSQL, RDS PostgreSQL, Aurora MySQL, RDS MySQL, and MariaDB)

db.application - The name of the application that is connected to the database (only Aurora PostgreSQL and RDS PostgreSQL)

db.host - The host name of the connected client (all engines)

db.session_type - The type of the current session (only Aurora PostgreSQL and RDS PostgreSQL)

db.sql - The SQL that is currently executing (all engines)

db.sql_tokenized - The SQL digest (all engines)

db.wait_event - The event for which the database backend is waiting (all engines)

db.wait_event_type - The type of event for which the database backend is waiting (all engines)

db.user - The user logged in to the database (all engines)

DimensionKeyDescription

dimensionKeyDescription_partitions :: Lens' DimensionKeyDescription (Maybe [Double])

If PartitionBy was specified, PartitionKeys contains the dimensions that were.

dimensionKeyDescription_total :: Lens' DimensionKeyDescription (Maybe Double)

The aggregated metric value for the dimension(s), over the requested time range.

dimensionKeyDescription_dimensions :: Lens' DimensionKeyDescription (Maybe (HashMap Text Text))

A map of name-value pairs for the dimensions in the group.

DimensionKeyDetail

dimensionKeyDetail_status :: Lens' DimensionKeyDetail (Maybe DetailStatus)

The status of the dimension detail data. Possible values include the following:

AVAILABLE - The dimension detail data is ready to be retrieved.

PROCESSING - The dimension detail data isn't ready to be retrieved because more processing time is required. If the requested detail data for db.sql.statement has the status PROCESSING, Performance Insights returns the truncated query.

UNAVAILABLE - The dimension detail data could not be collected successfully.

dimensionKeyDetail_dimension :: Lens' DimensionKeyDetail (Maybe Text)

The full name of the dimension. The full name includes the group name and key name. The only valid value is db.sql.statement.

dimensionKeyDetail_value :: Lens' DimensionKeyDetail (Maybe Text)

The value of the dimension detail data. For the db.sql.statement dimension, this value is either the full or truncated SQL query, depending on the return status.

MetricKeyDataPoints

metricKeyDataPoints_dataPoints :: Lens' MetricKeyDataPoints (Maybe [DataPoint])

An array of timestamp-value pairs, representing measurements over a period of time.

metricKeyDataPoints_key :: Lens' MetricKeyDataPoints (Maybe ResponseResourceMetricKey)

The dimension(s) to which the data points apply.

MetricQuery

metricQuery_groupBy :: Lens' MetricQuery (Maybe DimensionGroup)

A specification for how to aggregate the data points from a query result. You must specify a valid dimension group. Performance Insights will return all of the dimensions within that group, unless you provide the names of specific dimensions within that group. You can also request that Performance Insights return a limited number of values for a dimension.

metricQuery_filter :: Lens' MetricQuery (Maybe (HashMap Text Text))

One or more filters to apply in the request. Restrictions:

Any number of filters by the same dimension, as specified in the GroupBy parameter.

A single filter for any other dimension in this dimension group.

metricQuery_metric :: Lens' MetricQuery Text

The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

ResponsePartitionKey

responsePartitionKey_dimensions :: Lens' ResponsePartitionKey (HashMap Text Text)

A dimension map that contains the dimension(s) for this partition.

ResponseResourceMetricKey

responseResourceMetricKey_dimensions :: Lens' ResponseResourceMetricKey (Maybe (HashMap Text Text))

The valid dimensions for the metric.

responseResourceMetricKey_metric :: Lens' ResponseResourceMetricKey Text

The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.