Amazonka.Pi.Types

Description

Service Configuration

defaultService :: Service

API version 2018-02-27 of the Amazon Performance Insights SDK configuration.

Errors

_InvalidArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

One of the arguments provided is invalid for this request.

_NotAuthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

The user is not authorized to perform this request.

_InternalServiceError :: AsError a => Getting (First ServiceError) a ServiceError

The request failed due to an unknown error.

DetailStatus

newtype DetailStatus

Constructors

DetailStatus'

Fields

fromDetailStatus :: Text

Bundled Patterns

pattern DetailStatus_AVAILABLE :: DetailStatus

pattern DetailStatus_PROCESSING :: DetailStatus

pattern DetailStatus_UNAVAILABLE :: DetailStatus

ServiceType

newtype ServiceType

Constructors

ServiceType'

Fields

fromServiceType :: Text

Bundled Patterns

pattern ServiceType_RDS :: ServiceType

DataPoint

data DataPoint

A timestamp, and a single numerical value, which together represent a measurement at a particular point in time.

See: newDataPoint smart constructor.

Constructors

DataPoint'

Fields

timestamp :: POSIX

The time, in epoch format, associated with a particular Value.

value :: Double

The actual value associated with a particular Timestamp.

newDataPoint

Arguments

:: UTCTime
$sel:timestamp:DataPoint'

-> Double
$sel:value:DataPoint'

-> DataPoint

Create a value of DataPoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timestamp:DataPoint', dataPoint_timestamp - The time, in epoch format, associated with a particular Value.

$sel:value:DataPoint', dataPoint_value - The actual value associated with a particular Timestamp.

dataPoint_timestamp :: Lens' DataPoint UTCTime

The time, in epoch format, associated with a particular Value.

dataPoint_value :: Lens' DataPoint Double

The actual value associated with a particular Timestamp.

DimensionGroup

data DimensionGroup

A logical grouping of Performance Insights metrics for a related subject area. For example, the db.sql dimension group consists of the following dimensions: db.sql.id, db.sql.db_id, db.sql.statement, and db.sql.tokenized_id.

Each response element returns a maximum of 500 bytes. For larger elements, such as SQL statements, only the first 500 bytes are returned.

See: newDimensionGroup smart constructor.

Constructors

DimensionGroup'

Fields

limit :: Maybe Natural

The maximum number of items to fetch for this dimension group.

dimensions :: Maybe (NonEmpty Text)

A list of specific dimensions from a dimension group. If this parameter is not present, then it signifies that all of the dimensions in the group were requested, or are present in the response.

Valid values for elements in the Dimensions array are:

db.application.name - The name of the application that is connected to the database (only Aurora PostgreSQL and RDS PostgreSQL)

db.host.id - The host ID of the connected client (all engines)

db.host.name - The host name of the connected client (all engines)

db.name - The name of the database to which the client is connected (only Aurora PostgreSQL, RDS PostgreSQL, Aurora MySQL, RDS MySQL, and MariaDB)

db.session_type.name - The type of the current session (only Aurora PostgreSQL and RDS PostgreSQL)

db.sql.id - The SQL ID generated by Performance Insights (all engines)

db.sql.db_id - The SQL ID generated by the database (all engines)

db.sql.statement - The SQL text that is being executed (all engines)

db.sql.tokenized_id

db.sql_tokenized.id - The SQL digest ID generated by Performance Insights (all engines)

db.sql_tokenized.db_id - SQL digest ID generated by the database (all engines)

db.sql_tokenized.statement - The SQL digest text (all engines)

db.user.id - The ID of the user logged in to the database (all engines)

db.user.name - The name of the user logged in to the database (all engines)

db.wait_event.name - The event for which the backend is waiting (all engines)

db.wait_event.type - The type of event for which the backend is waiting (all engines)

db.wait_event_type.name - The name of the event type for which the backend is waiting (all engines)

group' :: Text

The name of the dimension group. Valid values are:

db - The name of the database to which the client is connected (only Aurora PostgreSQL, RDS PostgreSQL, Aurora MySQL, RDS MySQL, and MariaDB)

db.application - The name of the application that is connected to the database (only Aurora PostgreSQL and RDS PostgreSQL)

db.host - The host name of the connected client (all engines)

db.session_type - The type of the current session (only Aurora PostgreSQL and RDS PostgreSQL)

db.sql - The SQL that is currently executing (all engines)

db.sql_tokenized - The SQL digest (all engines)

db.wait_event - The event for which the database backend is waiting (all engines)

db.wait_event_type - The type of event for which the database backend is waiting (all engines)

db.user - The user logged in to the database (all engines)

newDimensionGroup

Arguments

:: Text
$sel:group':DimensionGroup'

-> DimensionGroup

Create a value of DimensionGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:limit:DimensionGroup', dimensionGroup_limit - The maximum number of items to fetch for this dimension group.

$sel:dimensions:DimensionGroup', dimensionGroup_dimensions - A list of specific dimensions from a dimension group. If this parameter is not present, then it signifies that all of the dimensions in the group were requested, or are present in the response.

Valid values for elements in the Dimensions array are:

db.application.name - The name of the application that is connected to the database (only Aurora PostgreSQL and RDS PostgreSQL)

db.host.id - The host ID of the connected client (all engines)

db.host.name - The host name of the connected client (all engines)

db.name - The name of the database to which the client is connected (only Aurora PostgreSQL, RDS PostgreSQL, Aurora MySQL, RDS MySQL, and MariaDB)

db.session_type.name - The type of the current session (only Aurora PostgreSQL and RDS PostgreSQL)

db.sql.id - The SQL ID generated by Performance Insights (all engines)

db.sql.db_id - The SQL ID generated by the database (all engines)

db.sql.statement - The SQL text that is being executed (all engines)

db.sql.tokenized_id

db.sql_tokenized.id - The SQL digest ID generated by Performance Insights (all engines)

db.sql_tokenized.db_id - SQL digest ID generated by the database (all engines)

db.sql_tokenized.statement - The SQL digest text (all engines)

db.user.id - The ID of the user logged in to the database (all engines)

db.user.name - The name of the user logged in to the database (all engines)

db.wait_event.name - The event for which the backend is waiting (all engines)

db.wait_event.type - The type of event for which the backend is waiting (all engines)

db.wait_event_type.name - The name of the event type for which the backend is waiting (all engines)

$sel:group':DimensionGroup', dimensionGroup_group - The name of the dimension group. Valid values are:

db - The name of the database to which the client is connected (only Aurora PostgreSQL, RDS PostgreSQL, Aurora MySQL, RDS MySQL, and MariaDB)

db.application - The name of the application that is connected to the database (only Aurora PostgreSQL and RDS PostgreSQL)

db.host - The host name of the connected client (all engines)

db.session_type - The type of the current session (only Aurora PostgreSQL and RDS PostgreSQL)

db.sql - The SQL that is currently executing (all engines)

db.sql_tokenized - The SQL digest (all engines)

db.wait_event - The event for which the database backend is waiting (all engines)

db.wait_event_type - The type of event for which the database backend is waiting (all engines)

db.user - The user logged in to the database (all engines)

dimensionGroup_limit :: Lens' DimensionGroup (Maybe Natural)

The maximum number of items to fetch for this dimension group.

dimensionGroup_dimensions :: Lens' DimensionGroup (Maybe (NonEmpty Text))

A list of specific dimensions from a dimension group. If this parameter is not present, then it signifies that all of the dimensions in the group were requested, or are present in the response.

Valid values for elements in the Dimensions array are:

db.application.name - The name of the application that is connected to the database (only Aurora PostgreSQL and RDS PostgreSQL)

db.host.id - The host ID of the connected client (all engines)

db.host.name - The host name of the connected client (all engines)

db.name - The name of the database to which the client is connected (only Aurora PostgreSQL, RDS PostgreSQL, Aurora MySQL, RDS MySQL, and MariaDB)

db.session_type.name - The type of the current session (only Aurora PostgreSQL and RDS PostgreSQL)

db.sql.id - The SQL ID generated by Performance Insights (all engines)

db.sql.db_id - The SQL ID generated by the database (all engines)

db.sql.statement - The SQL text that is being executed (all engines)

db.sql.tokenized_id

db.sql_tokenized.id - The SQL digest ID generated by Performance Insights (all engines)

db.sql_tokenized.db_id - SQL digest ID generated by the database (all engines)

db.sql_tokenized.statement - The SQL digest text (all engines)

db.user.id - The ID of the user logged in to the database (all engines)

db.user.name - The name of the user logged in to the database (all engines)

db.wait_event.name - The event for which the backend is waiting (all engines)

db.wait_event.type - The type of event for which the backend is waiting (all engines)

db.wait_event_type.name - The name of the event type for which the backend is waiting (all engines)

dimensionGroup_group :: Lens' DimensionGroup Text

The name of the dimension group. Valid values are:

db - The name of the database to which the client is connected (only Aurora PostgreSQL, RDS PostgreSQL, Aurora MySQL, RDS MySQL, and MariaDB)

db.application - The name of the application that is connected to the database (only Aurora PostgreSQL and RDS PostgreSQL)

db.host - The host name of the connected client (all engines)

db.session_type - The type of the current session (only Aurora PostgreSQL and RDS PostgreSQL)

db.sql - The SQL that is currently executing (all engines)

db.sql_tokenized - The SQL digest (all engines)

db.wait_event - The event for which the database backend is waiting (all engines)

db.wait_event_type - The type of event for which the database backend is waiting (all engines)

db.user - The user logged in to the database (all engines)

DimensionKeyDescription

data DimensionKeyDescription

An array of descriptions and aggregated values for each dimension within a dimension group.

See: newDimensionKeyDescription smart constructor.

Constructors

DimensionKeyDescription'

Fields

partitions :: Maybe [Double]

If PartitionBy was specified, PartitionKeys contains the dimensions that were.

total :: Maybe Double

The aggregated metric value for the dimension(s), over the requested time range.

dimensions :: Maybe (HashMap Text Text)

A map of name-value pairs for the dimensions in the group.

newDimensionKeyDescription :: DimensionKeyDescription

Create a value of DimensionKeyDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:partitions:DimensionKeyDescription', dimensionKeyDescription_partitions - If PartitionBy was specified, PartitionKeys contains the dimensions that were.

$sel:total:DimensionKeyDescription', dimensionKeyDescription_total - The aggregated metric value for the dimension(s), over the requested time range.

$sel:dimensions:DimensionKeyDescription', dimensionKeyDescription_dimensions - A map of name-value pairs for the dimensions in the group.

dimensionKeyDescription_partitions :: Lens' DimensionKeyDescription (Maybe [Double])

If PartitionBy was specified, PartitionKeys contains the dimensions that were.

dimensionKeyDescription_total :: Lens' DimensionKeyDescription (Maybe Double)

The aggregated metric value for the dimension(s), over the requested time range.

dimensionKeyDescription_dimensions :: Lens' DimensionKeyDescription (Maybe (HashMap Text Text))

A map of name-value pairs for the dimensions in the group.

DimensionKeyDetail

data DimensionKeyDetail

An object that describes the details for a specified dimension.

See: newDimensionKeyDetail smart constructor.

Constructors

DimensionKeyDetail'

Fields

status :: Maybe DetailStatus

The status of the dimension detail data. Possible values include the following:

AVAILABLE - The dimension detail data is ready to be retrieved.

PROCESSING - The dimension detail data isn't ready to be retrieved because more processing time is required. If the requested detail data for db.sql.statement has the status PROCESSING, Performance Insights returns the truncated query.

UNAVAILABLE - The dimension detail data could not be collected successfully.

dimension :: Maybe Text

The full name of the dimension. The full name includes the group name and key name. The only valid value is db.sql.statement.

value :: Maybe Text

The value of the dimension detail data. For the db.sql.statement dimension, this value is either the full or truncated SQL query, depending on the return status.

newDimensionKeyDetail :: DimensionKeyDetail

Create a value of DimensionKeyDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DimensionKeyDetail', dimensionKeyDetail_status - The status of the dimension detail data. Possible values include the following:

AVAILABLE - The dimension detail data is ready to be retrieved.

PROCESSING - The dimension detail data isn't ready to be retrieved because more processing time is required. If the requested detail data for db.sql.statement has the status PROCESSING, Performance Insights returns the truncated query.

UNAVAILABLE - The dimension detail data could not be collected successfully.

$sel:dimension:DimensionKeyDetail', dimensionKeyDetail_dimension - The full name of the dimension. The full name includes the group name and key name. The only valid value is db.sql.statement.

$sel:value:DimensionKeyDetail', dimensionKeyDetail_value - The value of the dimension detail data. For the db.sql.statement dimension, this value is either the full or truncated SQL query, depending on the return status.

dimensionKeyDetail_status :: Lens' DimensionKeyDetail (Maybe DetailStatus)

The status of the dimension detail data. Possible values include the following:

AVAILABLE - The dimension detail data is ready to be retrieved.

PROCESSING - The dimension detail data isn't ready to be retrieved because more processing time is required. If the requested detail data for db.sql.statement has the status PROCESSING, Performance Insights returns the truncated query.

UNAVAILABLE - The dimension detail data could not be collected successfully.

dimensionKeyDetail_dimension :: Lens' DimensionKeyDetail (Maybe Text)

The full name of the dimension. The full name includes the group name and key name. The only valid value is db.sql.statement.

dimensionKeyDetail_value :: Lens' DimensionKeyDetail (Maybe Text)

The value of the dimension detail data. For the db.sql.statement dimension, this value is either the full or truncated SQL query, depending on the return status.

MetricKeyDataPoints

data MetricKeyDataPoints

A time-ordered series of data points, corresponding to a dimension of a Performance Insights metric.

See: newMetricKeyDataPoints smart constructor.

Constructors

MetricKeyDataPoints'

Fields

dataPoints :: Maybe [DataPoint]

An array of timestamp-value pairs, representing measurements over a period of time.

key :: Maybe ResponseResourceMetricKey

The dimension(s) to which the data points apply.

newMetricKeyDataPoints :: MetricKeyDataPoints

Create a value of MetricKeyDataPoints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataPoints:MetricKeyDataPoints', metricKeyDataPoints_dataPoints - An array of timestamp-value pairs, representing measurements over a period of time.

$sel:key:MetricKeyDataPoints', metricKeyDataPoints_key - The dimension(s) to which the data points apply.

metricKeyDataPoints_dataPoints :: Lens' MetricKeyDataPoints (Maybe [DataPoint])

An array of timestamp-value pairs, representing measurements over a period of time.

metricKeyDataPoints_key :: Lens' MetricKeyDataPoints (Maybe ResponseResourceMetricKey)

The dimension(s) to which the data points apply.

MetricQuery

data MetricQuery

A single query to be processed. You must provide the metric to query. If no other parameters are specified, Performance Insights returns all of the data points for that metric. You can optionally request that the data points be aggregated by dimension group ( GroupBy), and return only those data points that match your criteria (Filter).

See: newMetricQuery smart constructor.

Constructors

MetricQuery'

Fields

groupBy :: Maybe DimensionGroup

A specification for how to aggregate the data points from a query result. You must specify a valid dimension group. Performance Insights will return all of the dimensions within that group, unless you provide the names of specific dimensions within that group. You can also request that Performance Insights return a limited number of values for a dimension.

filter' :: Maybe (HashMap Text Text)

One or more filters to apply in the request. Restrictions:

Any number of filters by the same dimension, as specified in the GroupBy parameter.

A single filter for any other dimension in this dimension group.

metric :: Text

The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

newMetricQuery

Arguments

:: Text
$sel:metric:MetricQuery'

-> MetricQuery

Create a value of MetricQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupBy:MetricQuery', metricQuery_groupBy - A specification for how to aggregate the data points from a query result. You must specify a valid dimension group. Performance Insights will return all of the dimensions within that group, unless you provide the names of specific dimensions within that group. You can also request that Performance Insights return a limited number of values for a dimension.

$sel:filter':MetricQuery', metricQuery_filter - One or more filters to apply in the request. Restrictions:

Any number of filters by the same dimension, as specified in the GroupBy parameter.

A single filter for any other dimension in this dimension group.

$sel:metric:MetricQuery', metricQuery_metric - The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

metricQuery_groupBy :: Lens' MetricQuery (Maybe DimensionGroup)

A specification for how to aggregate the data points from a query result. You must specify a valid dimension group. Performance Insights will return all of the dimensions within that group, unless you provide the names of specific dimensions within that group. You can also request that Performance Insights return a limited number of values for a dimension.

metricQuery_filter :: Lens' MetricQuery (Maybe (HashMap Text Text))

One or more filters to apply in the request. Restrictions:

Any number of filters by the same dimension, as specified in the GroupBy parameter.

A single filter for any other dimension in this dimension group.

metricQuery_metric :: Lens' MetricQuery Text

The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

ResponsePartitionKey

data ResponsePartitionKey

If PartitionBy was specified in a DescribeDimensionKeys request, the dimensions are returned in an array. Each element in the array specifies one dimension.

See: newResponsePartitionKey smart constructor.

Constructors

ResponsePartitionKey'

Fields

dimensions :: HashMap Text Text

A dimension map that contains the dimension(s) for this partition.

newResponsePartitionKey :: ResponsePartitionKey

Create a value of ResponsePartitionKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensions:ResponsePartitionKey', responsePartitionKey_dimensions - A dimension map that contains the dimension(s) for this partition.

responsePartitionKey_dimensions :: Lens' ResponsePartitionKey (HashMap Text Text)

A dimension map that contains the dimension(s) for this partition.

ResponseResourceMetricKey

data ResponseResourceMetricKey

An object describing a Performance Insights metric and one or more dimensions for that metric.

See: newResponseResourceMetricKey smart constructor.

Constructors

ResponseResourceMetricKey'

Fields

dimensions :: Maybe (HashMap Text Text)

The valid dimensions for the metric.

metric :: Text

The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

newResponseResourceMetricKey

Arguments

:: Text
$sel:metric:ResponseResourceMetricKey'

-> ResponseResourceMetricKey

Create a value of ResponseResourceMetricKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensions:ResponseResourceMetricKey', responseResourceMetricKey_dimensions - The valid dimensions for the metric.

$sel:metric:ResponseResourceMetricKey', responseResourceMetricKey_metric - The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

responseResourceMetricKey_dimensions :: Lens' ResponseResourceMetricKey (Maybe (HashMap Text Text))

The valid dimensions for the metric.

responseResourceMetricKey_metric :: Lens' ResponseResourceMetricKey Text

The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.