Amazonka.Pi.Types.DimensionKeyDescription

Description

Documentation

data DimensionKeyDescription

An array of descriptions and aggregated values for each dimension within a dimension group.

See: newDimensionKeyDescription smart constructor.

Constructors

DimensionKeyDescription'

Fields

partitions :: Maybe [Double]

If PartitionBy was specified, PartitionKeys contains the dimensions that were.

total :: Maybe Double

The aggregated metric value for the dimension(s), over the requested time range.

dimensions :: Maybe (HashMap Text Text)

A map of name-value pairs for the dimensions in the group.

newDimensionKeyDescription :: DimensionKeyDescription

Create a value of DimensionKeyDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:partitions:DimensionKeyDescription', dimensionKeyDescription_partitions - If PartitionBy was specified, PartitionKeys contains the dimensions that were.

$sel:total:DimensionKeyDescription', dimensionKeyDescription_total - The aggregated metric value for the dimension(s), over the requested time range.

$sel:dimensions:DimensionKeyDescription', dimensionKeyDescription_dimensions - A map of name-value pairs for the dimensions in the group.

dimensionKeyDescription_partitions :: Lens' DimensionKeyDescription (Maybe [Double])

If PartitionBy was specified, PartitionKeys contains the dimensions that were.

dimensionKeyDescription_total :: Lens' DimensionKeyDescription (Maybe Double)

The aggregated metric value for the dimension(s), over the requested time range.

dimensionKeyDescription_dimensions :: Lens' DimensionKeyDescription (Maybe (HashMap Text Text))

A map of name-value pairs for the dimensions in the group.