Amazonka.Pi.Types.ResponseResourceMetricKey

Description

Documentation

data ResponseResourceMetricKey

An object describing a Performance Insights metric and one or more dimensions for that metric.

See: newResponseResourceMetricKey smart constructor.

Constructors

ResponseResourceMetricKey'

Fields

dimensions :: Maybe (HashMap Text Text)

The valid dimensions for the metric.

metric :: Text

The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

newResponseResourceMetricKey

Arguments

:: Text
$sel:metric:ResponseResourceMetricKey'

-> ResponseResourceMetricKey

Create a value of ResponseResourceMetricKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensions:ResponseResourceMetricKey', responseResourceMetricKey_dimensions - The valid dimensions for the metric.

$sel:metric:ResponseResourceMetricKey', responseResourceMetricKey_metric - The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

responseResourceMetricKey_dimensions :: Lens' ResponseResourceMetricKey (Maybe (HashMap Text Text))

The valid dimensions for the metric.

responseResourceMetricKey_metric :: Lens' ResponseResourceMetricKey Text

The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.