Amazonka.Pinpoint.CreateEmailTemplate

Description

Creates a message template for messages that are sent through the email channel.

Creating a Request

data CreateEmailTemplate

See: newCreateEmailTemplate smart constructor.

Constructors

CreateEmailTemplate'

Fields

templateName :: Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

emailTemplateRequest :: EmailTemplateRequest

newCreateEmailTemplate

Arguments

:: Text
$sel:templateName:CreateEmailTemplate'

-> EmailTemplateRequest
$sel:emailTemplateRequest:CreateEmailTemplate'

-> CreateEmailTemplate

Create a value of CreateEmailTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:CreateEmailTemplate', createEmailTemplate_templateName - The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

$sel:emailTemplateRequest:CreateEmailTemplate', createEmailTemplate_emailTemplateRequest - Undocumented member.

Request Lenses

createEmailTemplate_templateName :: Lens' CreateEmailTemplate Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

createEmailTemplate_emailTemplateRequest :: Lens' CreateEmailTemplate EmailTemplateRequest

Undocumented member.

Destructuring the Response

data CreateEmailTemplateResponse

See: newCreateEmailTemplateResponse smart constructor.

Constructors

CreateEmailTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

createTemplateMessageBody :: CreateTemplateMessageBody

newCreateEmailTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateEmailTemplateResponse'

-> CreateTemplateMessageBody
$sel:createTemplateMessageBody:CreateEmailTemplateResponse'

-> CreateEmailTemplateResponse

Create a value of CreateEmailTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateEmailTemplateResponse', createEmailTemplateResponse_httpStatus - The response's http status code.

$sel:createTemplateMessageBody:CreateEmailTemplateResponse', createEmailTemplateResponse_createTemplateMessageBody - Undocumented member.

Response Lenses

createEmailTemplateResponse_httpStatus :: Lens' CreateEmailTemplateResponse Int

The response's http status code.

createEmailTemplateResponse_createTemplateMessageBody :: Lens' CreateEmailTemplateResponse CreateTemplateMessageBody

Undocumented member.