Amazonka.Pinpoint.CreatePushTemplate

Description

Creates a message template for messages that are sent through a push notification channel.

Creating a Request

data CreatePushTemplate

See: newCreatePushTemplate smart constructor.

Constructors

CreatePushTemplate'

Fields

templateName :: Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

pushNotificationTemplateRequest :: PushNotificationTemplateRequest

newCreatePushTemplate

Arguments

:: Text
$sel:templateName:CreatePushTemplate'

-> PushNotificationTemplateRequest
$sel:pushNotificationTemplateRequest:CreatePushTemplate'

-> CreatePushTemplate

Create a value of CreatePushTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:CreatePushTemplate', createPushTemplate_templateName - The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

$sel:pushNotificationTemplateRequest:CreatePushTemplate', createPushTemplate_pushNotificationTemplateRequest - Undocumented member.

Request Lenses

createPushTemplate_templateName :: Lens' CreatePushTemplate Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

createPushTemplate_pushNotificationTemplateRequest :: Lens' CreatePushTemplate PushNotificationTemplateRequest

Undocumented member.

Destructuring the Response

data CreatePushTemplateResponse

See: newCreatePushTemplateResponse smart constructor.

Constructors

CreatePushTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

createTemplateMessageBody :: CreateTemplateMessageBody

newCreatePushTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreatePushTemplateResponse'

-> CreateTemplateMessageBody
$sel:createTemplateMessageBody:CreatePushTemplateResponse'

-> CreatePushTemplateResponse

Create a value of CreatePushTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreatePushTemplateResponse', createPushTemplateResponse_httpStatus - The response's http status code.

$sel:createTemplateMessageBody:CreatePushTemplateResponse', createPushTemplateResponse_createTemplateMessageBody - Undocumented member.

Response Lenses

createPushTemplateResponse_httpStatus :: Lens' CreatePushTemplateResponse Int

The response's http status code.

createPushTemplateResponse_createTemplateMessageBody :: Lens' CreatePushTemplateResponse CreateTemplateMessageBody

Undocumented member.