Amazonka.Pinpoint.DeleteAdmChannel

Description

Disables the ADM channel for an application and deletes any existing settings for the channel.

Creating a Request

data DeleteAdmChannel

See: newDeleteAdmChannel smart constructor.

Constructors

DeleteAdmChannel'

Fields

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newDeleteAdmChannel

Arguments

:: Text
$sel:applicationId:DeleteAdmChannel'

-> DeleteAdmChannel

Create a value of DeleteAdmChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:DeleteAdmChannel', deleteAdmChannel_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

deleteAdmChannel_applicationId :: Lens' DeleteAdmChannel Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data DeleteAdmChannelResponse

See: newDeleteAdmChannelResponse smart constructor.

Constructors

DeleteAdmChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

aDMChannelResponse :: ADMChannelResponse

newDeleteAdmChannelResponse

Arguments

:: Int
$sel:httpStatus:DeleteAdmChannelResponse'

-> ADMChannelResponse
$sel:aDMChannelResponse:DeleteAdmChannelResponse'

-> DeleteAdmChannelResponse

Create a value of DeleteAdmChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteAdmChannelResponse', deleteAdmChannelResponse_httpStatus - The response's http status code.

$sel:aDMChannelResponse:DeleteAdmChannelResponse', deleteAdmChannelResponse_aDMChannelResponse - Undocumented member.

Response Lenses

deleteAdmChannelResponse_httpStatus :: Lens' DeleteAdmChannelResponse Int

The response's http status code.

deleteAdmChannelResponse_aDMChannelResponse :: Lens' DeleteAdmChannelResponse ADMChannelResponse

Undocumented member.