Amazonka.Pinpoint.DeleteVoiceChannel

Description

Disables the voice channel for an application and deletes any existing settings for the channel.

Creating a Request

data DeleteVoiceChannel

See: newDeleteVoiceChannel smart constructor.

Constructors

DeleteVoiceChannel'

Fields

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newDeleteVoiceChannel

Arguments

:: Text
$sel:applicationId:DeleteVoiceChannel'

-> DeleteVoiceChannel

Create a value of DeleteVoiceChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:DeleteVoiceChannel', deleteVoiceChannel_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

deleteVoiceChannel_applicationId :: Lens' DeleteVoiceChannel Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data DeleteVoiceChannelResponse

See: newDeleteVoiceChannelResponse smart constructor.

Constructors

DeleteVoiceChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

voiceChannelResponse :: VoiceChannelResponse

newDeleteVoiceChannelResponse

Arguments

:: Int
$sel:httpStatus:DeleteVoiceChannelResponse'

-> VoiceChannelResponse
$sel:voiceChannelResponse:DeleteVoiceChannelResponse'

-> DeleteVoiceChannelResponse

Create a value of DeleteVoiceChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteVoiceChannelResponse', deleteVoiceChannelResponse_httpStatus - The response's http status code.

$sel:voiceChannelResponse:DeleteVoiceChannelResponse', deleteVoiceChannelResponse_voiceChannelResponse - Undocumented member.

Response Lenses

deleteVoiceChannelResponse_httpStatus :: Lens' DeleteVoiceChannelResponse Int

The response's http status code.

deleteVoiceChannelResponse_voiceChannelResponse :: Lens' DeleteVoiceChannelResponse VoiceChannelResponse

Undocumented member.