Amazonka.Pinpoint.GetJourneyDateRangeKpi

Description

Retrieves (queries) pre-aggregated data for a standard engagement metric that applies to a journey.

Creating a Request

data GetJourneyDateRangeKpi

See: newGetJourneyDateRangeKpi smart constructor.

Constructors

GetJourneyDateRangeKpi'

Fields

startTime :: Maybe POSIX

The first date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-19T20:00:00Z for 8:00 PM UTC July 19, 2019. This value should also be fewer than 90 days from the current day.

nextToken :: Maybe Text

The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

endTime :: Maybe POSIX

The last date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-26T20:00:00Z for 8:00 PM UTC July 26, 2019.

pageSize :: Maybe Text

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

journeyId :: Text

The unique identifier for the journey.

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

kpiName :: Text

The name of the metric, also referred to as a /key performance indicator (KPI)/, to retrieve data for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. Examples are email-open-rate and successful-delivery-rate. For a list of valid values, see the Amazon Pinpoint Developer Guide.

newGetJourneyDateRangeKpi

Arguments

:: Text
$sel:journeyId:GetJourneyDateRangeKpi'

-> Text
$sel:applicationId:GetJourneyDateRangeKpi'

-> Text
$sel:kpiName:GetJourneyDateRangeKpi'

-> GetJourneyDateRangeKpi

Create a value of GetJourneyDateRangeKpi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:GetJourneyDateRangeKpi', getJourneyDateRangeKpi_startTime - The first date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-19T20:00:00Z for 8:00 PM UTC July 19, 2019. This value should also be fewer than 90 days from the current day.

$sel:nextToken:GetJourneyDateRangeKpi', getJourneyDateRangeKpi_nextToken - The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:endTime:GetJourneyDateRangeKpi', getJourneyDateRangeKpi_endTime - The last date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-26T20:00:00Z for 8:00 PM UTC July 26, 2019.

$sel:pageSize:GetJourneyDateRangeKpi', getJourneyDateRangeKpi_pageSize - The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:journeyId:GetJourneyDateRangeKpi', getJourneyDateRangeKpi_journeyId - The unique identifier for the journey.

$sel:applicationId:GetJourneyDateRangeKpi', getJourneyDateRangeKpi_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

$sel:kpiName:GetJourneyDateRangeKpi', getJourneyDateRangeKpi_kpiName - The name of the metric, also referred to as a /key performance indicator (KPI)/, to retrieve data for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. Examples are email-open-rate and successful-delivery-rate. For a list of valid values, see the Amazon Pinpoint Developer Guide.

Request Lenses

getJourneyDateRangeKpi_startTime :: Lens' GetJourneyDateRangeKpi (Maybe UTCTime)

The first date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-19T20:00:00Z for 8:00 PM UTC July 19, 2019. This value should also be fewer than 90 days from the current day.

getJourneyDateRangeKpi_nextToken :: Lens' GetJourneyDateRangeKpi (Maybe Text)

The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

getJourneyDateRangeKpi_endTime :: Lens' GetJourneyDateRangeKpi (Maybe UTCTime)

The last date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-26T20:00:00Z for 8:00 PM UTC July 26, 2019.

getJourneyDateRangeKpi_pageSize :: Lens' GetJourneyDateRangeKpi (Maybe Text)

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

getJourneyDateRangeKpi_journeyId :: Lens' GetJourneyDateRangeKpi Text

The unique identifier for the journey.

getJourneyDateRangeKpi_applicationId :: Lens' GetJourneyDateRangeKpi Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

getJourneyDateRangeKpi_kpiName :: Lens' GetJourneyDateRangeKpi Text

The name of the metric, also referred to as a /key performance indicator (KPI)/, to retrieve data for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. Examples are email-open-rate and successful-delivery-rate. For a list of valid values, see the Amazon Pinpoint Developer Guide.

Destructuring the Response

data GetJourneyDateRangeKpiResponse

See: newGetJourneyDateRangeKpiResponse smart constructor.

Constructors

GetJourneyDateRangeKpiResponse'

Fields

httpStatus :: Int

The response's http status code.

journeyDateRangeKpiResponse :: JourneyDateRangeKpiResponse

newGetJourneyDateRangeKpiResponse

Arguments

:: Int
$sel:httpStatus:GetJourneyDateRangeKpiResponse'

-> JourneyDateRangeKpiResponse
$sel:journeyDateRangeKpiResponse:GetJourneyDateRangeKpiResponse'

-> GetJourneyDateRangeKpiResponse

Create a value of GetJourneyDateRangeKpiResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetJourneyDateRangeKpiResponse', getJourneyDateRangeKpiResponse_httpStatus - The response's http status code.

$sel:journeyDateRangeKpiResponse:GetJourneyDateRangeKpiResponse', getJourneyDateRangeKpiResponse_journeyDateRangeKpiResponse - Undocumented member.

Response Lenses

getJourneyDateRangeKpiResponse_httpStatus :: Lens' GetJourneyDateRangeKpiResponse Int

The response's http status code.

getJourneyDateRangeKpiResponse_journeyDateRangeKpiResponse :: Lens' GetJourneyDateRangeKpiResponse JourneyDateRangeKpiResponse

Undocumented member.