Amazonka.Pinpoint.GetSegment

Description

Retrieves information about the configuration, dimension, and other settings for a specific segment that's associated with an application.

Creating a Request

data GetSegment

See: newGetSegment smart constructor.

Constructors

GetSegment'

Fields

segmentId :: Text

The unique identifier for the segment.

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newGetSegment

Arguments

:: Text
$sel:segmentId:GetSegment'

-> Text
$sel:applicationId:GetSegment'

-> GetSegment

Create a value of GetSegment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:segmentId:GetSegment', getSegment_segmentId - The unique identifier for the segment.

$sel:applicationId:GetSegment', getSegment_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

getSegment_segmentId :: Lens' GetSegment Text

The unique identifier for the segment.

getSegment_applicationId :: Lens' GetSegment Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data GetSegmentResponse

See: newGetSegmentResponse smart constructor.

Constructors

GetSegmentResponse'

Fields

httpStatus :: Int

The response's http status code.

segmentResponse :: SegmentResponse

newGetSegmentResponse

Arguments

:: Int
$sel:httpStatus:GetSegmentResponse'

-> SegmentResponse
$sel:segmentResponse:GetSegmentResponse'

-> GetSegmentResponse

Create a value of GetSegmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetSegmentResponse', getSegmentResponse_httpStatus - The response's http status code.

$sel:segmentResponse:GetSegmentResponse', getSegmentResponse_segmentResponse - Undocumented member.

Response Lenses

getSegmentResponse_httpStatus :: Lens' GetSegmentResponse Int

The response's http status code.

getSegmentResponse_segmentResponse :: Lens' GetSegmentResponse SegmentResponse

Undocumented member.