Amazonka.Pinpoint.GetSegmentImportJobs

Description

Retrieves information about the status and settings of the import jobs for a segment.

Creating a Request

data GetSegmentImportJobs

See: newGetSegmentImportJobs smart constructor.

Constructors

GetSegmentImportJobs'

Fields

token :: Maybe Text

The NextToken string that specifies which page of results to return in a paginated response.

pageSize :: Maybe Text

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

segmentId :: Text

The unique identifier for the segment.

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newGetSegmentImportJobs

Arguments

:: Text
$sel:segmentId:GetSegmentImportJobs'

-> Text
$sel:applicationId:GetSegmentImportJobs'

-> GetSegmentImportJobs

Create a value of GetSegmentImportJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:token:GetSegmentImportJobs', getSegmentImportJobs_token - The NextToken string that specifies which page of results to return in a paginated response.

$sel:pageSize:GetSegmentImportJobs', getSegmentImportJobs_pageSize - The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:segmentId:GetSegmentImportJobs', getSegmentImportJobs_segmentId - The unique identifier for the segment.

$sel:applicationId:GetSegmentImportJobs', getSegmentImportJobs_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

getSegmentImportJobs_token :: Lens' GetSegmentImportJobs (Maybe Text)

The NextToken string that specifies which page of results to return in a paginated response.

getSegmentImportJobs_pageSize :: Lens' GetSegmentImportJobs (Maybe Text)

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

getSegmentImportJobs_segmentId :: Lens' GetSegmentImportJobs Text

The unique identifier for the segment.

getSegmentImportJobs_applicationId :: Lens' GetSegmentImportJobs Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data GetSegmentImportJobsResponse

See: newGetSegmentImportJobsResponse smart constructor.

Constructors

GetSegmentImportJobsResponse'

Fields

httpStatus :: Int

The response's http status code.

importJobsResponse :: ImportJobsResponse

newGetSegmentImportJobsResponse

Arguments

:: Int
$sel:httpStatus:GetSegmentImportJobsResponse'

-> ImportJobsResponse
$sel:importJobsResponse:GetSegmentImportJobsResponse'

-> GetSegmentImportJobsResponse

Create a value of GetSegmentImportJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetSegmentImportJobsResponse', getSegmentImportJobsResponse_httpStatus - The response's http status code.

$sel:importJobsResponse:GetSegmentImportJobsResponse', getSegmentImportJobsResponse_importJobsResponse - Undocumented member.

Response Lenses

getSegmentImportJobsResponse_httpStatus :: Lens' GetSegmentImportJobsResponse Int

The response's http status code.

getSegmentImportJobsResponse_importJobsResponse :: Lens' GetSegmentImportJobsResponse ImportJobsResponse

Undocumented member.