Amazonka.Pinpoint.GetVoiceChannel

Description

Retrieves information about the status and settings of the voice channel for an application.

Creating a Request

data GetVoiceChannel

See: newGetVoiceChannel smart constructor.

Constructors

GetVoiceChannel'

Fields

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newGetVoiceChannel

Arguments

:: Text
$sel:applicationId:GetVoiceChannel'

-> GetVoiceChannel

Create a value of GetVoiceChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:GetVoiceChannel', getVoiceChannel_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

getVoiceChannel_applicationId :: Lens' GetVoiceChannel Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data GetVoiceChannelResponse

See: newGetVoiceChannelResponse smart constructor.

Constructors

GetVoiceChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

voiceChannelResponse :: VoiceChannelResponse

newGetVoiceChannelResponse

Arguments

:: Int
$sel:httpStatus:GetVoiceChannelResponse'

-> VoiceChannelResponse
$sel:voiceChannelResponse:GetVoiceChannelResponse'

-> GetVoiceChannelResponse

Create a value of GetVoiceChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetVoiceChannelResponse', getVoiceChannelResponse_httpStatus - The response's http status code.

$sel:voiceChannelResponse:GetVoiceChannelResponse', getVoiceChannelResponse_voiceChannelResponse - Undocumented member.

Response Lenses

getVoiceChannelResponse_httpStatus :: Lens' GetVoiceChannelResponse Int

The response's http status code.

getVoiceChannelResponse_voiceChannelResponse :: Lens' GetVoiceChannelResponse VoiceChannelResponse

Undocumented member.