Amazonka.Pinpoint.Types.ADMChannelResponse

Description

Documentation

data ADMChannelResponse

Provides information about the status and settings of the ADM (Amazon Device Messaging) channel for an application.

See: newADMChannelResponse smart constructor.

Constructors

ADMChannelResponse'

Fields

lastModifiedDate :: Maybe Text

The date and time when the ADM channel was last modified.

enabled :: Maybe Bool

Specifies whether the ADM channel is enabled for the application.

isArchived :: Maybe Bool

Specifies whether the ADM channel is archived.

applicationId :: Maybe Text

The unique identifier for the application that the ADM channel applies to.

version :: Maybe Int

The current version of the ADM channel.

id :: Maybe Text

(Deprecated) An identifier for the ADM channel. This property is retained only for backward compatibility.

creationDate :: Maybe Text

The date and time when the ADM channel was enabled.

lastModifiedBy :: Maybe Text

The user who last modified the ADM channel.

hasCredential :: Maybe Bool

(Not used) This property is retained only for backward compatibility.

platform :: Text

The type of messaging or notification platform for the channel. For the ADM channel, this value is ADM.

newADMChannelResponse

Arguments

:: Text
$sel:platform:ADMChannelResponse'

-> ADMChannelResponse

Create a value of ADMChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:ADMChannelResponse', aDMChannelResponse_lastModifiedDate - The date and time when the ADM channel was last modified.

$sel:enabled:ADMChannelResponse', aDMChannelResponse_enabled - Specifies whether the ADM channel is enabled for the application.

$sel:isArchived:ADMChannelResponse', aDMChannelResponse_isArchived - Specifies whether the ADM channel is archived.

$sel:applicationId:ADMChannelResponse', aDMChannelResponse_applicationId - The unique identifier for the application that the ADM channel applies to.

$sel:version:ADMChannelResponse', aDMChannelResponse_version - The current version of the ADM channel.

$sel:id:ADMChannelResponse', aDMChannelResponse_id - (Deprecated) An identifier for the ADM channel. This property is retained only for backward compatibility.

$sel:creationDate:ADMChannelResponse', aDMChannelResponse_creationDate - The date and time when the ADM channel was enabled.

$sel:lastModifiedBy:ADMChannelResponse', aDMChannelResponse_lastModifiedBy - The user who last modified the ADM channel.

$sel:hasCredential:ADMChannelResponse', aDMChannelResponse_hasCredential - (Not used) This property is retained only for backward compatibility.

$sel:platform:ADMChannelResponse', aDMChannelResponse_platform - The type of messaging or notification platform for the channel. For the ADM channel, this value is ADM.

aDMChannelResponse_lastModifiedDate :: Lens' ADMChannelResponse (Maybe Text)

The date and time when the ADM channel was last modified.

aDMChannelResponse_enabled :: Lens' ADMChannelResponse (Maybe Bool)

Specifies whether the ADM channel is enabled for the application.

aDMChannelResponse_isArchived :: Lens' ADMChannelResponse (Maybe Bool)

Specifies whether the ADM channel is archived.

aDMChannelResponse_applicationId :: Lens' ADMChannelResponse (Maybe Text)

The unique identifier for the application that the ADM channel applies to.

aDMChannelResponse_version :: Lens' ADMChannelResponse (Maybe Int)

The current version of the ADM channel.

aDMChannelResponse_id :: Lens' ADMChannelResponse (Maybe Text)

(Deprecated) An identifier for the ADM channel. This property is retained only for backward compatibility.

aDMChannelResponse_creationDate :: Lens' ADMChannelResponse (Maybe Text)

The date and time when the ADM channel was enabled.

aDMChannelResponse_lastModifiedBy :: Lens' ADMChannelResponse (Maybe Text)

The user who last modified the ADM channel.

aDMChannelResponse_hasCredential :: Lens' ADMChannelResponse (Maybe Bool)

(Not used) This property is retained only for backward compatibility.

aDMChannelResponse_platform :: Lens' ADMChannelResponse Text

The type of messaging or notification platform for the channel. For the ADM channel, this value is ADM.