Amazonka.Pinpoint.Types.ADMMessage

Description

Documentation

data ADMMessage

Specifies the settings for a one-time message that's sent directly to an endpoint through the ADM (Amazon Device Messaging) channel.

See: newADMMessage smart constructor.

Constructors

ADMMessage'

Fields

substitutions :: Maybe (HashMap Text [Text])

The default message variables to use in the notification message. You can override the default variables with individual address variables.

expiresAfter :: Maybe Text

The amount of time, in seconds, that ADM should store the message if the recipient's device is offline. Amazon Pinpoint specifies this value in the expiresAfter parameter when it sends the notification message to ADM.

md5 :: Maybe Text

The base64-encoded, MD5 checksum of the value specified by the Data property. ADM uses the MD5 value to verify the integrity of the data.

silentPush :: Maybe Bool

Specifies whether the notification is a silent push notification, which is a push notification that doesn't display on a recipient's device. Silent push notifications can be used for cases such as updating an app's configuration or supporting phone home functionality.

imageIconUrl :: Maybe Text

The URL of the large icon image to display in the content view of the push notification.

rawContent :: Maybe Text

The raw, JSON-formatted string to use as the payload for the notification message. If specified, this value overrides all other content for the message.

data' :: Maybe (HashMap Text Text)

The JSON data payload to use for the push notification, if the notification is a silent push notification. This payload is added to the data.pinpoint.jsonBody object of the notification.

smallImageIconUrl :: Maybe Text

The URL of the small icon image to display in the status bar and the content view of the push notification.

body :: Maybe Text

The body of the notification message.

url :: Maybe Text

The URL to open in the recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.

sound :: Maybe Text

The sound to play when the recipient receives the push notification. You can use the default stream or specify the file name of a sound resource that's bundled in your app. On an Android platform, the sound file must reside in /res/raw/.

action :: Maybe Action

The action to occur if the recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This action uses the deep-linking features of the Android platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

imageUrl :: Maybe Text

The URL of an image to display in the push notification.

consolidationKey :: Maybe Text

An arbitrary string that indicates that multiple messages are logically the same and that Amazon Device Messaging (ADM) can drop previously enqueued messages in favor of this message.

title :: Maybe Text

The title to display above the notification message on the recipient's device.

iconReference :: Maybe Text

The icon image name of the asset saved in your app.

newADMMessage :: ADMMessage

Create a value of ADMMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:substitutions:ADMMessage', aDMMessage_substitutions - The default message variables to use in the notification message. You can override the default variables with individual address variables.

$sel:expiresAfter:ADMMessage', aDMMessage_expiresAfter - The amount of time, in seconds, that ADM should store the message if the recipient's device is offline. Amazon Pinpoint specifies this value in the expiresAfter parameter when it sends the notification message to ADM.

$sel:md5:ADMMessage', aDMMessage_md5 - The base64-encoded, MD5 checksum of the value specified by the Data property. ADM uses the MD5 value to verify the integrity of the data.

$sel:silentPush:ADMMessage', aDMMessage_silentPush - Specifies whether the notification is a silent push notification, which is a push notification that doesn't display on a recipient's device. Silent push notifications can be used for cases such as updating an app's configuration or supporting phone home functionality.

$sel:imageIconUrl:ADMMessage', aDMMessage_imageIconUrl - The URL of the large icon image to display in the content view of the push notification.

$sel:rawContent:ADMMessage', aDMMessage_rawContent - The raw, JSON-formatted string to use as the payload for the notification message. If specified, this value overrides all other content for the message.

$sel:data':ADMMessage', aDMMessage_data - The JSON data payload to use for the push notification, if the notification is a silent push notification. This payload is added to the data.pinpoint.jsonBody object of the notification.

$sel:smallImageIconUrl:ADMMessage', aDMMessage_smallImageIconUrl - The URL of the small icon image to display in the status bar and the content view of the push notification.

$sel:body:ADMMessage', aDMMessage_body - The body of the notification message.

$sel:url:ADMMessage', aDMMessage_url - The URL to open in the recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.

$sel:sound:ADMMessage', aDMMessage_sound - The sound to play when the recipient receives the push notification. You can use the default stream or specify the file name of a sound resource that's bundled in your app. On an Android platform, the sound file must reside in /res/raw/.

$sel:action:ADMMessage', aDMMessage_action - The action to occur if the recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This action uses the deep-linking features of the Android platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

$sel:imageUrl:ADMMessage', aDMMessage_imageUrl - The URL of an image to display in the push notification.

$sel:consolidationKey:ADMMessage', aDMMessage_consolidationKey - An arbitrary string that indicates that multiple messages are logically the same and that Amazon Device Messaging (ADM) can drop previously enqueued messages in favor of this message.

$sel:title:ADMMessage', aDMMessage_title - The title to display above the notification message on the recipient's device.

$sel:iconReference:ADMMessage', aDMMessage_iconReference - The icon image name of the asset saved in your app.

aDMMessage_substitutions :: Lens' ADMMessage (Maybe (HashMap Text [Text]))

The default message variables to use in the notification message. You can override the default variables with individual address variables.

aDMMessage_expiresAfter :: Lens' ADMMessage (Maybe Text)

The amount of time, in seconds, that ADM should store the message if the recipient's device is offline. Amazon Pinpoint specifies this value in the expiresAfter parameter when it sends the notification message to ADM.

aDMMessage_md5 :: Lens' ADMMessage (Maybe Text)

The base64-encoded, MD5 checksum of the value specified by the Data property. ADM uses the MD5 value to verify the integrity of the data.

aDMMessage_silentPush :: Lens' ADMMessage (Maybe Bool)

Specifies whether the notification is a silent push notification, which is a push notification that doesn't display on a recipient's device. Silent push notifications can be used for cases such as updating an app's configuration or supporting phone home functionality.

aDMMessage_imageIconUrl :: Lens' ADMMessage (Maybe Text)

The URL of the large icon image to display in the content view of the push notification.

aDMMessage_rawContent :: Lens' ADMMessage (Maybe Text)

The raw, JSON-formatted string to use as the payload for the notification message. If specified, this value overrides all other content for the message.

aDMMessage_data :: Lens' ADMMessage (Maybe (HashMap Text Text))

The JSON data payload to use for the push notification, if the notification is a silent push notification. This payload is added to the data.pinpoint.jsonBody object of the notification.

aDMMessage_smallImageIconUrl :: Lens' ADMMessage (Maybe Text)

The URL of the small icon image to display in the status bar and the content view of the push notification.

aDMMessage_body :: Lens' ADMMessage (Maybe Text)

The body of the notification message.

aDMMessage_url :: Lens' ADMMessage (Maybe Text)

The URL to open in the recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.

aDMMessage_sound :: Lens' ADMMessage (Maybe Text)

The sound to play when the recipient receives the push notification. You can use the default stream or specify the file name of a sound resource that's bundled in your app. On an Android platform, the sound file must reside in /res/raw/.

aDMMessage_action :: Lens' ADMMessage (Maybe Action)

The action to occur if the recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This action uses the deep-linking features of the Android platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

aDMMessage_imageUrl :: Lens' ADMMessage (Maybe Text)

The URL of an image to display in the push notification.

aDMMessage_consolidationKey :: Lens' ADMMessage (Maybe Text)

An arbitrary string that indicates that multiple messages are logically the same and that Amazon Device Messaging (ADM) can drop previously enqueued messages in favor of this message.

aDMMessage_title :: Lens' ADMMessage (Maybe Text)

The title to display above the notification message on the recipient's device.

aDMMessage_iconReference :: Lens' ADMMessage (Maybe Text)

The icon image name of the asset saved in your app.