Amazonka.Pinpoint.Types.APNSChannelResponse

Description

Documentation

data APNSChannelResponse

Provides information about the status and settings of the APNs (Apple Push Notification service) channel for an application.

See: newAPNSChannelResponse smart constructor.

Constructors

APNSChannelResponse'

Fields

lastModifiedDate :: Maybe Text

The date and time when the APNs channel was last modified.

enabled :: Maybe Bool

Specifies whether the APNs channel is enabled for the application.

hasTokenKey :: Maybe Bool

Specifies whether the APNs channel is configured to communicate with APNs by using APNs tokens. To provide an authentication key for APNs tokens, set the TokenKey property of the channel.

defaultAuthenticationMethod :: Maybe Text

The default authentication method that Amazon Pinpoint uses to authenticate with APNs for this channel, key or certificate.

isArchived :: Maybe Bool

Specifies whether the APNs channel is archived.

applicationId :: Maybe Text

The unique identifier for the application that the APNs channel applies to.

version :: Maybe Int

The current version of the APNs channel.

id :: Maybe Text

(Deprecated) An identifier for the APNs channel. This property is retained only for backward compatibility.

creationDate :: Maybe Text

The date and time when the APNs channel was enabled.

lastModifiedBy :: Maybe Text

The user who last modified the APNs channel.

hasCredential :: Maybe Bool

(Not used) This property is retained only for backward compatibility.

platform :: Text

The type of messaging or notification platform for the channel. For the APNs channel, this value is APNS.

newAPNSChannelResponse

Arguments

:: Text
$sel:platform:APNSChannelResponse'

-> APNSChannelResponse

Create a value of APNSChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:APNSChannelResponse', aPNSChannelResponse_lastModifiedDate - The date and time when the APNs channel was last modified.

$sel:enabled:APNSChannelResponse', aPNSChannelResponse_enabled - Specifies whether the APNs channel is enabled for the application.

$sel:hasTokenKey:APNSChannelResponse', aPNSChannelResponse_hasTokenKey - Specifies whether the APNs channel is configured to communicate with APNs by using APNs tokens. To provide an authentication key for APNs tokens, set the TokenKey property of the channel.

$sel:defaultAuthenticationMethod:APNSChannelResponse', aPNSChannelResponse_defaultAuthenticationMethod - The default authentication method that Amazon Pinpoint uses to authenticate with APNs for this channel, key or certificate.

$sel:isArchived:APNSChannelResponse', aPNSChannelResponse_isArchived - Specifies whether the APNs channel is archived.

$sel:applicationId:APNSChannelResponse', aPNSChannelResponse_applicationId - The unique identifier for the application that the APNs channel applies to.

$sel:version:APNSChannelResponse', aPNSChannelResponse_version - The current version of the APNs channel.

$sel:id:APNSChannelResponse', aPNSChannelResponse_id - (Deprecated) An identifier for the APNs channel. This property is retained only for backward compatibility.

$sel:creationDate:APNSChannelResponse', aPNSChannelResponse_creationDate - The date and time when the APNs channel was enabled.

$sel:lastModifiedBy:APNSChannelResponse', aPNSChannelResponse_lastModifiedBy - The user who last modified the APNs channel.

$sel:hasCredential:APNSChannelResponse', aPNSChannelResponse_hasCredential - (Not used) This property is retained only for backward compatibility.

$sel:platform:APNSChannelResponse', aPNSChannelResponse_platform - The type of messaging or notification platform for the channel. For the APNs channel, this value is APNS.

aPNSChannelResponse_lastModifiedDate :: Lens' APNSChannelResponse (Maybe Text)

The date and time when the APNs channel was last modified.

aPNSChannelResponse_enabled :: Lens' APNSChannelResponse (Maybe Bool)

Specifies whether the APNs channel is enabled for the application.

aPNSChannelResponse_hasTokenKey :: Lens' APNSChannelResponse (Maybe Bool)

Specifies whether the APNs channel is configured to communicate with APNs by using APNs tokens. To provide an authentication key for APNs tokens, set the TokenKey property of the channel.

aPNSChannelResponse_defaultAuthenticationMethod :: Lens' APNSChannelResponse (Maybe Text)

The default authentication method that Amazon Pinpoint uses to authenticate with APNs for this channel, key or certificate.

aPNSChannelResponse_isArchived :: Lens' APNSChannelResponse (Maybe Bool)

Specifies whether the APNs channel is archived.

aPNSChannelResponse_applicationId :: Lens' APNSChannelResponse (Maybe Text)

The unique identifier for the application that the APNs channel applies to.

aPNSChannelResponse_version :: Lens' APNSChannelResponse (Maybe Int)

The current version of the APNs channel.

aPNSChannelResponse_id :: Lens' APNSChannelResponse (Maybe Text)

(Deprecated) An identifier for the APNs channel. This property is retained only for backward compatibility.

aPNSChannelResponse_creationDate :: Lens' APNSChannelResponse (Maybe Text)

The date and time when the APNs channel was enabled.

aPNSChannelResponse_lastModifiedBy :: Lens' APNSChannelResponse (Maybe Text)

The user who last modified the APNs channel.

aPNSChannelResponse_hasCredential :: Lens' APNSChannelResponse (Maybe Bool)

(Not used) This property is retained only for backward compatibility.

aPNSChannelResponse_platform :: Lens' APNSChannelResponse Text

The type of messaging or notification platform for the channel. For the APNs channel, this value is APNS.