Amazonka.Pinpoint.Types.ActivityResponse

Description

Documentation

data ActivityResponse

Provides information about an activity that was performed by a campaign.

See: newActivityResponse smart constructor.

Constructors

ActivityResponse'

Fields

state :: Maybe Text

The current status of the activity. Possible values are: PENDING, INITIALIZING, RUNNING, PAUSED, CANCELLED, and COMPLETED.

start :: Maybe Text

The actual start time, in ISO 8601 format, of the activity.

timezonesCompletedCount :: Maybe Int

The total number of time zones that were completed.

timezonesTotalCount :: Maybe Int

The total number of unique time zones that are in the segment for the campaign.

result :: Maybe Text

Specifies whether the activity succeeded. Possible values are SUCCESS and FAIL.

treatmentId :: Maybe Text

The unique identifier for the campaign treatment that the activity applies to. A treatment is a variation of a campaign that's used for A/B testing of a campaign.

successfulEndpointCount :: Maybe Int

The total number of endpoints that the campaign successfully delivered messages to.

end :: Maybe Text

The actual time, in ISO 8601 format, when the activity was marked CANCELLED or COMPLETED.

totalEndpointCount :: Maybe Int

The total number of endpoints that the campaign attempted to deliver messages to.

scheduledStart :: Maybe Text

The scheduled start time, in ISO 8601 format, for the activity.

campaignId :: Text

The unique identifier for the campaign that the activity applies to.

id :: Text

The unique identifier for the activity.

applicationId :: Text

The unique identifier for the application that the campaign applies to.

newActivityResponse

Arguments

:: Text
$sel:campaignId:ActivityResponse'

-> Text
$sel:id:ActivityResponse'

-> Text
$sel:applicationId:ActivityResponse'

-> ActivityResponse

Create a value of ActivityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ActivityResponse', activityResponse_state - The current status of the activity. Possible values are: PENDING, INITIALIZING, RUNNING, PAUSED, CANCELLED, and COMPLETED.

$sel:start:ActivityResponse', activityResponse_start - The actual start time, in ISO 8601 format, of the activity.

$sel:timezonesCompletedCount:ActivityResponse', activityResponse_timezonesCompletedCount - The total number of time zones that were completed.

$sel:timezonesTotalCount:ActivityResponse', activityResponse_timezonesTotalCount - The total number of unique time zones that are in the segment for the campaign.

$sel:result:ActivityResponse', activityResponse_result - Specifies whether the activity succeeded. Possible values are SUCCESS and FAIL.

$sel:treatmentId:ActivityResponse', activityResponse_treatmentId - The unique identifier for the campaign treatment that the activity applies to. A treatment is a variation of a campaign that's used for A/B testing of a campaign.

$sel:successfulEndpointCount:ActivityResponse', activityResponse_successfulEndpointCount - The total number of endpoints that the campaign successfully delivered messages to.

$sel:end:ActivityResponse', activityResponse_end - The actual time, in ISO 8601 format, when the activity was marked CANCELLED or COMPLETED.

$sel:totalEndpointCount:ActivityResponse', activityResponse_totalEndpointCount - The total number of endpoints that the campaign attempted to deliver messages to.

$sel:scheduledStart:ActivityResponse', activityResponse_scheduledStart - The scheduled start time, in ISO 8601 format, for the activity.

$sel:campaignId:ActivityResponse', activityResponse_campaignId - The unique identifier for the campaign that the activity applies to.

$sel:id:ActivityResponse', activityResponse_id - The unique identifier for the activity.

$sel:applicationId:ActivityResponse', activityResponse_applicationId - The unique identifier for the application that the campaign applies to.

activityResponse_state :: Lens' ActivityResponse (Maybe Text)

The current status of the activity. Possible values are: PENDING, INITIALIZING, RUNNING, PAUSED, CANCELLED, and COMPLETED.

activityResponse_start :: Lens' ActivityResponse (Maybe Text)

The actual start time, in ISO 8601 format, of the activity.

activityResponse_timezonesCompletedCount :: Lens' ActivityResponse (Maybe Int)

The total number of time zones that were completed.

activityResponse_timezonesTotalCount :: Lens' ActivityResponse (Maybe Int)

The total number of unique time zones that are in the segment for the campaign.

activityResponse_result :: Lens' ActivityResponse (Maybe Text)

Specifies whether the activity succeeded. Possible values are SUCCESS and FAIL.

activityResponse_treatmentId :: Lens' ActivityResponse (Maybe Text)

The unique identifier for the campaign treatment that the activity applies to. A treatment is a variation of a campaign that's used for A/B testing of a campaign.

activityResponse_successfulEndpointCount :: Lens' ActivityResponse (Maybe Int)

The total number of endpoints that the campaign successfully delivered messages to.

activityResponse_end :: Lens' ActivityResponse (Maybe Text)

The actual time, in ISO 8601 format, when the activity was marked CANCELLED or COMPLETED.

activityResponse_totalEndpointCount :: Lens' ActivityResponse (Maybe Int)

The total number of endpoints that the campaign attempted to deliver messages to.

activityResponse_scheduledStart :: Lens' ActivityResponse (Maybe Text)

The scheduled start time, in ISO 8601 format, for the activity.

activityResponse_campaignId :: Lens' ActivityResponse Text

The unique identifier for the campaign that the activity applies to.

activityResponse_id :: Lens' ActivityResponse Text

The unique identifier for the activity.

activityResponse_applicationId :: Lens' ActivityResponse Text

The unique identifier for the application that the campaign applies to.