Amazonka.Pinpoint.Types.AttributeDimension

Description

Documentation

data AttributeDimension

Specifies attribute-based criteria for including or excluding endpoints from a segment.

See: newAttributeDimension smart constructor.

Constructors

AttributeDimension'

Fields

attributeType :: Maybe AttributeType

The type of segment dimension to use. Valid values are:

INCLUSIVE - endpoints that have attributes matching the values are included in the segment.

EXCLUSIVE - endpoints that have attributes matching the values are excluded in the segment.

CONTAINS - endpoints that have attributes' substrings match the values are included in the segment.

BEFORE - endpoints with attributes read as ISO_INSTANT datetimes before the value are included in the segment.

AFTER - endpoints with attributes read as ISO_INSTANT datetimes after the value are included in the segment.

ON - endpoints with attributes read as ISO_INSTANT dates on the value are included in the segment. Time is ignored in this comparison.

BETWEEN - endpoints with attributes read as ISO_INSTANT datetimes between the values are included in the segment.

values :: [Text]

The criteria values to use for the segment dimension. Depending on the value of the AttributeType property, endpoints are included or excluded from the segment if their attribute values match the criteria values.

newAttributeDimension :: AttributeDimension

Create a value of AttributeDimension with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeType:AttributeDimension', attributeDimension_attributeType - The type of segment dimension to use. Valid values are:

INCLUSIVE - endpoints that have attributes matching the values are included in the segment.

EXCLUSIVE - endpoints that have attributes matching the values are excluded in the segment.

CONTAINS - endpoints that have attributes' substrings match the values are included in the segment.

BEFORE - endpoints with attributes read as ISO_INSTANT datetimes before the value are included in the segment.

AFTER - endpoints with attributes read as ISO_INSTANT datetimes after the value are included in the segment.

ON - endpoints with attributes read as ISO_INSTANT dates on the value are included in the segment. Time is ignored in this comparison.

BETWEEN - endpoints with attributes read as ISO_INSTANT datetimes between the values are included in the segment.

$sel:values:AttributeDimension', attributeDimension_values - The criteria values to use for the segment dimension. Depending on the value of the AttributeType property, endpoints are included or excluded from the segment if their attribute values match the criteria values.

attributeDimension_attributeType :: Lens' AttributeDimension (Maybe AttributeType)

The type of segment dimension to use. Valid values are:

INCLUSIVE - endpoints that have attributes matching the values are included in the segment.

EXCLUSIVE - endpoints that have attributes matching the values are excluded in the segment.

CONTAINS - endpoints that have attributes' substrings match the values are included in the segment.

BEFORE - endpoints with attributes read as ISO_INSTANT datetimes before the value are included in the segment.

AFTER - endpoints with attributes read as ISO_INSTANT datetimes after the value are included in the segment.

ON - endpoints with attributes read as ISO_INSTANT dates on the value are included in the segment. Time is ignored in this comparison.

BETWEEN - endpoints with attributes read as ISO_INSTANT datetimes between the values are included in the segment.

attributeDimension_values :: Lens' AttributeDimension [Text]

The criteria values to use for the segment dimension. Depending on the value of the AttributeType property, endpoints are included or excluded from the segment if their attribute values match the criteria values.