Amazonka.Pinpoint.Types.CampaignDateRangeKpiResponse

Description

Documentation

data CampaignDateRangeKpiResponse

Provides the results of a query that retrieved the data for a standard metric that applies to a campaign, and provides information about that query.

See: newCampaignDateRangeKpiResponse smart constructor.

Constructors

CampaignDateRangeKpiResponse'

Fields

nextToken :: Maybe Text

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null for the Campaign Metrics resource because the resource returns all results in a single page.

kpiResult :: BaseKpiResult

An array of objects that contains the results of the query. Each object contains the value for the metric and metadata about that value.

kpiName :: Text

The name of the metric, also referred to as a /key performance indicator (KPI)/, that the data was retrieved for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. For a list of possible values, see the Amazon Pinpoint Developer Guide.

endTime :: POSIX

The last date and time of the date range that was used to filter the query results, in extended ISO 8601 format. The date range is inclusive.

campaignId :: Text

The unique identifier for the campaign that the metric applies to.

startTime :: POSIX

The first date and time of the date range that was used to filter the query results, in extended ISO 8601 format. The date range is inclusive.

applicationId :: Text

The unique identifier for the application that the metric applies to.

newCampaignDateRangeKpiResponse

Arguments

:: BaseKpiResult
$sel:kpiResult:CampaignDateRangeKpiResponse'

-> Text
$sel:kpiName:CampaignDateRangeKpiResponse'

-> UTCTime
$sel:endTime:CampaignDateRangeKpiResponse'

-> Text
$sel:campaignId:CampaignDateRangeKpiResponse'

-> UTCTime
$sel:startTime:CampaignDateRangeKpiResponse'

-> Text
$sel:applicationId:CampaignDateRangeKpiResponse'

-> CampaignDateRangeKpiResponse

Create a value of CampaignDateRangeKpiResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:CampaignDateRangeKpiResponse', campaignDateRangeKpiResponse_nextToken - The string to use in a subsequent request to get the next page of results in a paginated response. This value is null for the Campaign Metrics resource because the resource returns all results in a single page.

$sel:kpiResult:CampaignDateRangeKpiResponse', campaignDateRangeKpiResponse_kpiResult - An array of objects that contains the results of the query. Each object contains the value for the metric and metadata about that value.

$sel:kpiName:CampaignDateRangeKpiResponse', campaignDateRangeKpiResponse_kpiName - The name of the metric, also referred to as a /key performance indicator (KPI)/, that the data was retrieved for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. For a list of possible values, see the Amazon Pinpoint Developer Guide.

$sel:endTime:CampaignDateRangeKpiResponse', campaignDateRangeKpiResponse_endTime - The last date and time of the date range that was used to filter the query results, in extended ISO 8601 format. The date range is inclusive.

$sel:campaignId:CampaignDateRangeKpiResponse', campaignDateRangeKpiResponse_campaignId - The unique identifier for the campaign that the metric applies to.

$sel:startTime:CampaignDateRangeKpiResponse', campaignDateRangeKpiResponse_startTime - The first date and time of the date range that was used to filter the query results, in extended ISO 8601 format. The date range is inclusive.

$sel:applicationId:CampaignDateRangeKpiResponse', campaignDateRangeKpiResponse_applicationId - The unique identifier for the application that the metric applies to.

campaignDateRangeKpiResponse_nextToken :: Lens' CampaignDateRangeKpiResponse (Maybe Text)

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null for the Campaign Metrics resource because the resource returns all results in a single page.

campaignDateRangeKpiResponse_kpiResult :: Lens' CampaignDateRangeKpiResponse BaseKpiResult

An array of objects that contains the results of the query. Each object contains the value for the metric and metadata about that value.

campaignDateRangeKpiResponse_kpiName :: Lens' CampaignDateRangeKpiResponse Text

The name of the metric, also referred to as a /key performance indicator (KPI)/, that the data was retrieved for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. For a list of possible values, see the Amazon Pinpoint Developer Guide.

campaignDateRangeKpiResponse_endTime :: Lens' CampaignDateRangeKpiResponse UTCTime

The last date and time of the date range that was used to filter the query results, in extended ISO 8601 format. The date range is inclusive.

campaignDateRangeKpiResponse_campaignId :: Lens' CampaignDateRangeKpiResponse Text

The unique identifier for the campaign that the metric applies to.

campaignDateRangeKpiResponse_startTime :: Lens' CampaignDateRangeKpiResponse UTCTime

The first date and time of the date range that was used to filter the query results, in extended ISO 8601 format. The date range is inclusive.

campaignDateRangeKpiResponse_applicationId :: Lens' CampaignDateRangeKpiResponse Text

The unique identifier for the application that the metric applies to.