Amazonka.Pinpoint.Types.CampaignEmailMessage

Description

Documentation

data CampaignEmailMessage

Specifies the content and "From" address for an email message that's sent to recipients of a campaign.

See: newCampaignEmailMessage smart constructor.

Constructors

CampaignEmailMessage'

Fields

body :: Maybe Text

The body of the email for recipients whose email clients don't render HTML content.

fromAddress :: Maybe Text

The verified email address to send the email from. The default address is the FromAddress specified for the email channel for the application.

htmlBody :: Maybe Text

The body of the email, in HTML format, for recipients whose email clients render HTML content.

title :: Maybe Text

The subject line, or title, of the email.

newCampaignEmailMessage :: CampaignEmailMessage

Create a value of CampaignEmailMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:body:CampaignEmailMessage', campaignEmailMessage_body - The body of the email for recipients whose email clients don't render HTML content.

$sel:fromAddress:CampaignEmailMessage', campaignEmailMessage_fromAddress - The verified email address to send the email from. The default address is the FromAddress specified for the email channel for the application.

$sel:htmlBody:CampaignEmailMessage', campaignEmailMessage_htmlBody - The body of the email, in HTML format, for recipients whose email clients render HTML content.

$sel:title:CampaignEmailMessage', campaignEmailMessage_title - The subject line, or title, of the email.

campaignEmailMessage_body :: Lens' CampaignEmailMessage (Maybe Text)

The body of the email for recipients whose email clients don't render HTML content.

campaignEmailMessage_fromAddress :: Lens' CampaignEmailMessage (Maybe Text)

The verified email address to send the email from. The default address is the FromAddress specified for the email channel for the application.

campaignEmailMessage_htmlBody :: Lens' CampaignEmailMessage (Maybe Text)

The body of the email, in HTML format, for recipients whose email clients render HTML content.

campaignEmailMessage_title :: Lens' CampaignEmailMessage (Maybe Text)

The subject line, or title, of the email.