Amazonka.Pinpoint.Types.CampaignInAppMessage

Description

Documentation

data CampaignInAppMessage

In-app message configuration.

See: newCampaignInAppMessage smart constructor.

Constructors

CampaignInAppMessage'

Fields

layout :: Maybe Layout

In-app message layout.

body :: Maybe Text

The message body of the notification, the email body or the text message.

content :: Maybe [InAppMessageContent]

In-app message content.

customConfig :: Maybe (HashMap Text Text)

Custom config to be sent to client.

newCampaignInAppMessage :: CampaignInAppMessage

Create a value of CampaignInAppMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:layout:CampaignInAppMessage', campaignInAppMessage_layout - In-app message layout.

$sel:body:CampaignInAppMessage', campaignInAppMessage_body - The message body of the notification, the email body or the text message.

$sel:content:CampaignInAppMessage', campaignInAppMessage_content - In-app message content.

$sel:customConfig:CampaignInAppMessage', campaignInAppMessage_customConfig - Custom config to be sent to client.

campaignInAppMessage_layout :: Lens' CampaignInAppMessage (Maybe Layout)

In-app message layout.

campaignInAppMessage_body :: Lens' CampaignInAppMessage (Maybe Text)

The message body of the notification, the email body or the text message.

campaignInAppMessage_content :: Lens' CampaignInAppMessage (Maybe [InAppMessageContent])

In-app message content.

campaignInAppMessage_customConfig :: Lens' CampaignInAppMessage (Maybe (HashMap Text Text))

Custom config to be sent to client.