Amazonka.Pinpoint.Types.ChannelResponse

Description

Documentation

data ChannelResponse

Provides information about the general settings and status of a channel for an application.

See: newChannelResponse smart constructor.

Constructors

ChannelResponse'

Fields

lastModifiedDate :: Maybe Text

The date and time, in ISO 8601 format, when the channel was last modified.

enabled :: Maybe Bool

Specifies whether the channel is enabled for the application.

isArchived :: Maybe Bool

Specifies whether the channel is archived.

applicationId :: Maybe Text

The unique identifier for the application.

version :: Maybe Int

The current version of the channel.

id :: Maybe Text

(Deprecated) An identifier for the channel. This property is retained only for backward compatibility.

creationDate :: Maybe Text

The date and time, in ISO 8601 format, when the channel was enabled.

lastModifiedBy :: Maybe Text

The user who last modified the channel.

hasCredential :: Maybe Bool

(Not used) This property is retained only for backward compatibility.

newChannelResponse :: ChannelResponse

Create a value of ChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:ChannelResponse', channelResponse_lastModifiedDate - The date and time, in ISO 8601 format, when the channel was last modified.

$sel:enabled:ChannelResponse', channelResponse_enabled - Specifies whether the channel is enabled for the application.

$sel:isArchived:ChannelResponse', channelResponse_isArchived - Specifies whether the channel is archived.

$sel:applicationId:ChannelResponse', channelResponse_applicationId - The unique identifier for the application.

$sel:version:ChannelResponse', channelResponse_version - The current version of the channel.

$sel:id:ChannelResponse', channelResponse_id - (Deprecated) An identifier for the channel. This property is retained only for backward compatibility.

$sel:creationDate:ChannelResponse', channelResponse_creationDate - The date and time, in ISO 8601 format, when the channel was enabled.

$sel:lastModifiedBy:ChannelResponse', channelResponse_lastModifiedBy - The user who last modified the channel.

$sel:hasCredential:ChannelResponse', channelResponse_hasCredential - (Not used) This property is retained only for backward compatibility.

channelResponse_lastModifiedDate :: Lens' ChannelResponse (Maybe Text)

The date and time, in ISO 8601 format, when the channel was last modified.

channelResponse_enabled :: Lens' ChannelResponse (Maybe Bool)

Specifies whether the channel is enabled for the application.

channelResponse_isArchived :: Lens' ChannelResponse (Maybe Bool)

Specifies whether the channel is archived.

channelResponse_applicationId :: Lens' ChannelResponse (Maybe Text)

The unique identifier for the application.

channelResponse_version :: Lens' ChannelResponse (Maybe Int)

The current version of the channel.

channelResponse_id :: Lens' ChannelResponse (Maybe Text)

(Deprecated) An identifier for the channel. This property is retained only for backward compatibility.

channelResponse_creationDate :: Lens' ChannelResponse (Maybe Text)

The date and time, in ISO 8601 format, when the channel was enabled.

channelResponse_lastModifiedBy :: Lens' ChannelResponse (Maybe Text)

The user who last modified the channel.

channelResponse_hasCredential :: Lens' ChannelResponse (Maybe Bool)

(Not used) This property is retained only for backward compatibility.