Amazonka.Pinpoint.Types.EmailMessage

Description

Documentation

data EmailMessage

Specifies the default settings and content for a one-time email message that's sent directly to an endpoint.

See: newEmailMessage smart constructor.

Constructors

EmailMessage'

Fields

substitutions :: Maybe (HashMap Text [Text])

The default message variables to use in the email message. You can override the default variables with individual address variables.

body :: Maybe Text

The body of the email message.

fromAddress :: Maybe Text

The verified email address to send the email message from. The default value is the FromAddress specified for the email channel.

rawEmail :: Maybe RawEmail

The email message, represented as a raw MIME message.

feedbackForwardingAddress :: Maybe Text

The email address to forward bounces and complaints to, if feedback forwarding is enabled.

simpleEmail :: Maybe SimpleEmail

The email message, composed of a subject, a text part, and an HTML part.

replyToAddresses :: Maybe [Text]

The reply-to email address(es) for the email message. If a recipient replies to the email, each reply-to address receives the reply.

newEmailMessage :: EmailMessage

Create a value of EmailMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:substitutions:EmailMessage', emailMessage_substitutions - The default message variables to use in the email message. You can override the default variables with individual address variables.

$sel:body:EmailMessage', emailMessage_body - The body of the email message.

$sel:fromAddress:EmailMessage', emailMessage_fromAddress - The verified email address to send the email message from. The default value is the FromAddress specified for the email channel.

$sel:rawEmail:EmailMessage', emailMessage_rawEmail - The email message, represented as a raw MIME message.

$sel:feedbackForwardingAddress:EmailMessage', emailMessage_feedbackForwardingAddress - The email address to forward bounces and complaints to, if feedback forwarding is enabled.

$sel:simpleEmail:EmailMessage', emailMessage_simpleEmail - The email message, composed of a subject, a text part, and an HTML part.

$sel:replyToAddresses:EmailMessage', emailMessage_replyToAddresses - The reply-to email address(es) for the email message. If a recipient replies to the email, each reply-to address receives the reply.

emailMessage_substitutions :: Lens' EmailMessage (Maybe (HashMap Text [Text]))

The default message variables to use in the email message. You can override the default variables with individual address variables.

emailMessage_body :: Lens' EmailMessage (Maybe Text)

The body of the email message.

emailMessage_fromAddress :: Lens' EmailMessage (Maybe Text)

The verified email address to send the email message from. The default value is the FromAddress specified for the email channel.

emailMessage_rawEmail :: Lens' EmailMessage (Maybe RawEmail)

The email message, represented as a raw MIME message.

emailMessage_feedbackForwardingAddress :: Lens' EmailMessage (Maybe Text)

The email address to forward bounces and complaints to, if feedback forwarding is enabled.

emailMessage_simpleEmail :: Lens' EmailMessage (Maybe SimpleEmail)

The email message, composed of a subject, a text part, and an HTML part.

emailMessage_replyToAddresses :: Lens' EmailMessage (Maybe [Text])

The reply-to email address(es) for the email message. If a recipient replies to the email, each reply-to address receives the reply.