Amazonka.Pinpoint.Types.EmailMessageActivity

Description

Documentation

data EmailMessageActivity

Specifies the settings for an email activity in a journey. This type of activity sends an email message to participants.

See: newEmailMessageActivity smart constructor.

Constructors

EmailMessageActivity'

Fields

templateName :: Maybe Text

The name of the email message template to use for the message. If specified, this value must match the name of an existing message template.

templateVersion :: Maybe Text

The unique identifier for the version of the email template to use for the message. If specified, this value must match the identifier for an existing template version. To retrieve a list of versions and version identifiers for a template, use the Template Versions resource.

If you don't specify a value for this property, Amazon Pinpoint uses the active version of the template. The active version is typically the version of a template that's been most recently reviewed and approved for use, depending on your workflow. It isn't necessarily the latest version of a template.

nextActivity :: Maybe Text

The unique identifier for the next activity to perform, after the message is sent.

messageConfig :: Maybe JourneyEmailMessage

Specifies the sender address for an email message that's sent to participants in the journey.

newEmailMessageActivity :: EmailMessageActivity

Create a value of EmailMessageActivity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:EmailMessageActivity', emailMessageActivity_templateName - The name of the email message template to use for the message. If specified, this value must match the name of an existing message template.

$sel:templateVersion:EmailMessageActivity', emailMessageActivity_templateVersion - The unique identifier for the version of the email template to use for the message. If specified, this value must match the identifier for an existing template version. To retrieve a list of versions and version identifiers for a template, use the Template Versions resource.

If you don't specify a value for this property, Amazon Pinpoint uses the active version of the template. The active version is typically the version of a template that's been most recently reviewed and approved for use, depending on your workflow. It isn't necessarily the latest version of a template.

$sel:nextActivity:EmailMessageActivity', emailMessageActivity_nextActivity - The unique identifier for the next activity to perform, after the message is sent.

$sel:messageConfig:EmailMessageActivity', emailMessageActivity_messageConfig - Specifies the sender address for an email message that's sent to participants in the journey.

emailMessageActivity_templateName :: Lens' EmailMessageActivity (Maybe Text)

The name of the email message template to use for the message. If specified, this value must match the name of an existing message template.

emailMessageActivity_templateVersion :: Lens' EmailMessageActivity (Maybe Text)

The unique identifier for the version of the email template to use for the message. If specified, this value must match the identifier for an existing template version. To retrieve a list of versions and version identifiers for a template, use the Template Versions resource.

If you don't specify a value for this property, Amazon Pinpoint uses the active version of the template. The active version is typically the version of a template that's been most recently reviewed and approved for use, depending on your workflow. It isn't necessarily the latest version of a template.

emailMessageActivity_nextActivity :: Lens' EmailMessageActivity (Maybe Text)

The unique identifier for the next activity to perform, after the message is sent.

emailMessageActivity_messageConfig :: Lens' EmailMessageActivity (Maybe JourneyEmailMessage)

Specifies the sender address for an email message that's sent to participants in the journey.