Amazonka.Pinpoint.Types.EmailTemplateResponse

Description

Documentation

data EmailTemplateResponse

Provides information about the content and settings for a message template that can be used in messages that are sent through the email channel.

See: newEmailTemplateResponse smart constructor.

Constructors

EmailTemplateResponse'

Fields

subject :: Maybe Text

The subject line, or title, that's used in email messages that are based on the message template.

textPart :: Maybe Text

The message body, in plain text format, that's used in email messages that are based on the message template.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the message template.

templateDescription :: Maybe Text

The custom description of the message template.

defaultSubstitutions :: Maybe Text

The JSON object that specifies the default values that are used for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

version :: Maybe Text

The unique identifier, as an integer, for the active version of the message template, or the version of the template that you specified by using the version parameter in your request.

htmlPart :: Maybe Text

The message body, in HTML format, that's used in email messages that are based on the message template.

recommenderId :: Maybe Text

The unique identifier for the recommender model that's used by the message template.

tags :: Maybe (HashMap Text Text)

A string-to-string map of key-value pairs that identifies the tags that are associated with the message template. Each tag consists of a required tag key and an associated tag value.

lastModifiedDate :: Text

The date, in ISO 8601 format, when the message template was last modified.

creationDate :: Text

The date, in ISO 8601 format, when the message template was created.

templateName :: Text

The name of the message template.

templateType :: TemplateType

The type of channel that the message template is designed for. For an email template, this value is EMAIL.

newEmailTemplateResponse

Arguments

:: Text
$sel:lastModifiedDate:EmailTemplateResponse'

-> Text
$sel:creationDate:EmailTemplateResponse'

-> Text
$sel:templateName:EmailTemplateResponse'

-> TemplateType
$sel:templateType:EmailTemplateResponse'

-> EmailTemplateResponse

Create a value of EmailTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subject:EmailTemplateResponse', emailTemplateResponse_subject - The subject line, or title, that's used in email messages that are based on the message template.

$sel:textPart:EmailTemplateResponse', emailTemplateResponse_textPart - The message body, in plain text format, that's used in email messages that are based on the message template.

$sel:arn:EmailTemplateResponse', emailTemplateResponse_arn - The Amazon Resource Name (ARN) of the message template.

$sel:templateDescription:EmailTemplateResponse', emailTemplateResponse_templateDescription - The custom description of the message template.

$sel:defaultSubstitutions:EmailTemplateResponse', emailTemplateResponse_defaultSubstitutions - The JSON object that specifies the default values that are used for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

$sel:version:EmailTemplateResponse', emailTemplateResponse_version - The unique identifier, as an integer, for the active version of the message template, or the version of the template that you specified by using the version parameter in your request.

$sel:htmlPart:EmailTemplateResponse', emailTemplateResponse_htmlPart - The message body, in HTML format, that's used in email messages that are based on the message template.

$sel:recommenderId:EmailTemplateResponse', emailTemplateResponse_recommenderId - The unique identifier for the recommender model that's used by the message template.

$sel:tags:EmailTemplateResponse', emailTemplateResponse_tags - A string-to-string map of key-value pairs that identifies the tags that are associated with the message template. Each tag consists of a required tag key and an associated tag value.

$sel:lastModifiedDate:EmailTemplateResponse', emailTemplateResponse_lastModifiedDate - The date, in ISO 8601 format, when the message template was last modified.

$sel:creationDate:EmailTemplateResponse', emailTemplateResponse_creationDate - The date, in ISO 8601 format, when the message template was created.

$sel:templateName:EmailTemplateResponse', emailTemplateResponse_templateName - The name of the message template.

$sel:templateType:EmailTemplateResponse', emailTemplateResponse_templateType - The type of channel that the message template is designed for. For an email template, this value is EMAIL.

emailTemplateResponse_subject :: Lens' EmailTemplateResponse (Maybe Text)

The subject line, or title, that's used in email messages that are based on the message template.

emailTemplateResponse_textPart :: Lens' EmailTemplateResponse (Maybe Text)

The message body, in plain text format, that's used in email messages that are based on the message template.

emailTemplateResponse_arn :: Lens' EmailTemplateResponse (Maybe Text)

The Amazon Resource Name (ARN) of the message template.

emailTemplateResponse_templateDescription :: Lens' EmailTemplateResponse (Maybe Text)

The custom description of the message template.

emailTemplateResponse_defaultSubstitutions :: Lens' EmailTemplateResponse (Maybe Text)

The JSON object that specifies the default values that are used for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

emailTemplateResponse_version :: Lens' EmailTemplateResponse (Maybe Text)

The unique identifier, as an integer, for the active version of the message template, or the version of the template that you specified by using the version parameter in your request.

emailTemplateResponse_htmlPart :: Lens' EmailTemplateResponse (Maybe Text)

The message body, in HTML format, that's used in email messages that are based on the message template.

emailTemplateResponse_recommenderId :: Lens' EmailTemplateResponse (Maybe Text)

The unique identifier for the recommender model that's used by the message template.

emailTemplateResponse_tags :: Lens' EmailTemplateResponse (Maybe (HashMap Text Text))

A string-to-string map of key-value pairs that identifies the tags that are associated with the message template. Each tag consists of a required tag key and an associated tag value.

emailTemplateResponse_lastModifiedDate :: Lens' EmailTemplateResponse Text

The date, in ISO 8601 format, when the message template was last modified.

emailTemplateResponse_creationDate :: Lens' EmailTemplateResponse Text

The date, in ISO 8601 format, when the message template was created.

emailTemplateResponse_templateName :: Lens' EmailTemplateResponse Text

The name of the message template.

emailTemplateResponse_templateType :: Lens' EmailTemplateResponse TemplateType

The type of channel that the message template is designed for. For an email template, this value is EMAIL.