Amazonka.Pinpoint.Types.EndpointLocation

Description

Documentation

data EndpointLocation

Specifies geographic information about an endpoint.

See: newEndpointLocation smart constructor.

Constructors

EndpointLocation'

Fields

postalCode :: Maybe Text

The postal or ZIP code for the area where the endpoint is located.

latitude :: Maybe Double

The latitude coordinate of the endpoint location, rounded to one decimal place.

country :: Maybe Text

The two-character code, in ISO 3166-1 alpha-2 format, for the country or region where the endpoint is located. For example, US for the United States.

city :: Maybe Text

The name of the city where the endpoint is located.

region :: Maybe Text

The name of the region where the endpoint is located. For locations in the United States, this value is the name of a state.

longitude :: Maybe Double

The longitude coordinate of the endpoint location, rounded to one decimal place.

newEndpointLocation :: EndpointLocation

Create a value of EndpointLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:postalCode:EndpointLocation', endpointLocation_postalCode - The postal or ZIP code for the area where the endpoint is located.

$sel:latitude:EndpointLocation', endpointLocation_latitude - The latitude coordinate of the endpoint location, rounded to one decimal place.

$sel:country:EndpointLocation', endpointLocation_country - The two-character code, in ISO 3166-1 alpha-2 format, for the country or region where the endpoint is located. For example, US for the United States.

$sel:city:EndpointLocation', endpointLocation_city - The name of the city where the endpoint is located.

$sel:region:EndpointLocation', endpointLocation_region - The name of the region where the endpoint is located. For locations in the United States, this value is the name of a state.

$sel:longitude:EndpointLocation', endpointLocation_longitude - The longitude coordinate of the endpoint location, rounded to one decimal place.

endpointLocation_postalCode :: Lens' EndpointLocation (Maybe Text)

The postal or ZIP code for the area where the endpoint is located.

endpointLocation_latitude :: Lens' EndpointLocation (Maybe Double)

The latitude coordinate of the endpoint location, rounded to one decimal place.

endpointLocation_country :: Lens' EndpointLocation (Maybe Text)

The two-character code, in ISO 3166-1 alpha-2 format, for the country or region where the endpoint is located. For example, US for the United States.

endpointLocation_city :: Lens' EndpointLocation (Maybe Text)

The name of the city where the endpoint is located.

endpointLocation_region :: Lens' EndpointLocation (Maybe Text)

The name of the region where the endpoint is located. For locations in the United States, this value is the name of a state.

endpointLocation_longitude :: Lens' EndpointLocation (Maybe Double)

The longitude coordinate of the endpoint location, rounded to one decimal place.