Amazonka.Pinpoint.Types.EndpointMessageResult

Description

Documentation

data EndpointMessageResult

Provides information about the delivery status and results of sending a message directly to an endpoint.

See: newEndpointMessageResult smart constructor.

Constructors

EndpointMessageResult'

Fields

address :: Maybe Text

The endpoint address that the message was delivered to.

statusMessage :: Maybe Text

The status message for delivering the message.

updatedToken :: Maybe Text

For push notifications that are sent through the GCM channel, specifies whether the endpoint's device registration token was updated as part of delivering the message.

messageId :: Maybe Text

The unique identifier for the message that was sent.

deliveryStatus :: DeliveryStatus

The delivery status of the message. Possible values are:

DUPLICATE - The endpoint address is a duplicate of another endpoint address. Amazon Pinpoint won't attempt to send the message again.

OPT_OUT - The user who's associated with the endpoint has opted out of receiving messages from you. Amazon Pinpoint won't attempt to send the message again.

PERMANENT_FAILURE - An error occurred when delivering the message to the endpoint. Amazon Pinpoint won't attempt to send the message again.

SUCCESSFUL - The message was successfully delivered to the endpoint.

TEMPORARY_FAILURE - A temporary error occurred. Amazon Pinpoint won't attempt to send the message again.

THROTTLED - Amazon Pinpoint throttled the operation to send the message to the endpoint.

TIMEOUT - The message couldn't be sent within the timeout period.

UNKNOWN_FAILURE - An unknown error occurred.

statusCode :: Int

The downstream service status code for delivering the message.

newEndpointMessageResult

Arguments

:: DeliveryStatus
$sel:deliveryStatus:EndpointMessageResult'

-> Int
$sel:statusCode:EndpointMessageResult'

-> EndpointMessageResult

Create a value of EndpointMessageResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:address:EndpointMessageResult', endpointMessageResult_address - The endpoint address that the message was delivered to.

$sel:statusMessage:EndpointMessageResult', endpointMessageResult_statusMessage - The status message for delivering the message.

$sel:updatedToken:EndpointMessageResult', endpointMessageResult_updatedToken - For push notifications that are sent through the GCM channel, specifies whether the endpoint's device registration token was updated as part of delivering the message.

$sel:messageId:EndpointMessageResult', endpointMessageResult_messageId - The unique identifier for the message that was sent.

$sel:deliveryStatus:EndpointMessageResult', endpointMessageResult_deliveryStatus - The delivery status of the message. Possible values are:

DUPLICATE - The endpoint address is a duplicate of another endpoint address. Amazon Pinpoint won't attempt to send the message again.

OPT_OUT - The user who's associated with the endpoint has opted out of receiving messages from you. Amazon Pinpoint won't attempt to send the message again.

PERMANENT_FAILURE - An error occurred when delivering the message to the endpoint. Amazon Pinpoint won't attempt to send the message again.

SUCCESSFUL - The message was successfully delivered to the endpoint.

TEMPORARY_FAILURE - A temporary error occurred. Amazon Pinpoint won't attempt to send the message again.

THROTTLED - Amazon Pinpoint throttled the operation to send the message to the endpoint.

TIMEOUT - The message couldn't be sent within the timeout period.

UNKNOWN_FAILURE - An unknown error occurred.

$sel:statusCode:EndpointMessageResult', endpointMessageResult_statusCode - The downstream service status code for delivering the message.

endpointMessageResult_address :: Lens' EndpointMessageResult (Maybe Text)

The endpoint address that the message was delivered to.

endpointMessageResult_statusMessage :: Lens' EndpointMessageResult (Maybe Text)

The status message for delivering the message.

endpointMessageResult_updatedToken :: Lens' EndpointMessageResult (Maybe Text)

For push notifications that are sent through the GCM channel, specifies whether the endpoint's device registration token was updated as part of delivering the message.

endpointMessageResult_messageId :: Lens' EndpointMessageResult (Maybe Text)

The unique identifier for the message that was sent.

endpointMessageResult_deliveryStatus :: Lens' EndpointMessageResult DeliveryStatus

The delivery status of the message. Possible values are:

DUPLICATE - The endpoint address is a duplicate of another endpoint address. Amazon Pinpoint won't attempt to send the message again.

OPT_OUT - The user who's associated with the endpoint has opted out of receiving messages from you. Amazon Pinpoint won't attempt to send the message again.

PERMANENT_FAILURE - An error occurred when delivering the message to the endpoint. Amazon Pinpoint won't attempt to send the message again.

SUCCESSFUL - The message was successfully delivered to the endpoint.

TEMPORARY_FAILURE - A temporary error occurred. Amazon Pinpoint won't attempt to send the message again.

THROTTLED - Amazon Pinpoint throttled the operation to send the message to the endpoint.

TIMEOUT - The message couldn't be sent within the timeout period.

UNKNOWN_FAILURE - An unknown error occurred.

endpointMessageResult_statusCode :: Lens' EndpointMessageResult Int

The downstream service status code for delivering the message.