Amazonka.Pinpoint.Types.EndpointRequest

Description

Documentation

data EndpointRequest

Specifies the channel type and other settings for an endpoint.

See: newEndpointRequest smart constructor.

Constructors

EndpointRequest'

Fields

requestId :: Maybe Text

The unique identifier for the most recent request to update the endpoint.

metrics :: Maybe (HashMap Text Double)

One or more custom metrics that your app reports to Amazon Pinpoint for the endpoint.

location :: Maybe EndpointLocation

The geographic information for the endpoint.

demographic :: Maybe EndpointDemographic

The demographic information for the endpoint, such as the time zone and platform.

address :: Maybe Text

The destination address for messages or push notifications that you send to the endpoint. The address varies by channel. For a push-notification channel, use the token provided by the push notification service, such as an Apple Push Notification service (APNs) device token or a Firebase Cloud Messaging (FCM) registration token. For the SMS channel, use a phone number in E.164 format, such as +12065550100. For the email channel, use an email address.

effectiveDate :: Maybe Text

The date and time, in ISO 8601 format, when the endpoint is updated.

user :: Maybe EndpointUser

One or more custom attributes that describe the user who's associated with the endpoint.

attributes :: Maybe (HashMap Text [Text])

One or more custom attributes that describe the endpoint by associating a name with an array of values. For example, the value of a custom attribute named Interests might be: ["Science", "Music", "Travel"]. You can use these attributes as filter criteria when you create segments. Attribute names are case sensitive.

An attribute name can contain up to 50 characters. An attribute value can contain up to 100 characters. When you define the name of a custom attribute, avoid using the following characters: number sign (#), colon (:), question mark (?), backslash (\), and slash (/). The Amazon Pinpoint console can't display attribute names that contain these characters. This restriction doesn't apply to attribute values.

endpointStatus :: Maybe Text

Specifies whether to send messages or push notifications to the endpoint. Valid values are: ACTIVE, messages are sent to the endpoint; and, INACTIVE, messages aren’t sent to the endpoint.

Amazon Pinpoint automatically sets this value to ACTIVE when you create an endpoint or update an existing endpoint. Amazon Pinpoint automatically sets this value to INACTIVE if you update another endpoint that has the same address specified by the Address property.

optOut :: Maybe Text

Specifies whether the user who's associated with the endpoint has opted out of receiving messages and push notifications from you. Possible values are: ALL, the user has opted out and doesn't want to receive any messages or push notifications; and, NONE, the user hasn't opted out and wants to receive all messages and push notifications.

channelType :: Maybe ChannelType

The channel to use when sending messages or push notifications to the endpoint.

newEndpointRequest :: EndpointRequest

Create a value of EndpointRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:EndpointRequest', endpointRequest_requestId - The unique identifier for the most recent request to update the endpoint.

$sel:metrics:EndpointRequest', endpointRequest_metrics - One or more custom metrics that your app reports to Amazon Pinpoint for the endpoint.

$sel:location:EndpointRequest', endpointRequest_location - The geographic information for the endpoint.

$sel:demographic:EndpointRequest', endpointRequest_demographic - The demographic information for the endpoint, such as the time zone and platform.

$sel:address:EndpointRequest', endpointRequest_address - The destination address for messages or push notifications that you send to the endpoint. The address varies by channel. For a push-notification channel, use the token provided by the push notification service, such as an Apple Push Notification service (APNs) device token or a Firebase Cloud Messaging (FCM) registration token. For the SMS channel, use a phone number in E.164 format, such as +12065550100. For the email channel, use an email address.

$sel:effectiveDate:EndpointRequest', endpointRequest_effectiveDate - The date and time, in ISO 8601 format, when the endpoint is updated.

$sel:user:EndpointRequest', endpointRequest_user - One or more custom attributes that describe the user who's associated with the endpoint.

$sel:attributes:EndpointRequest', endpointRequest_attributes - One or more custom attributes that describe the endpoint by associating a name with an array of values. For example, the value of a custom attribute named Interests might be: ["Science", "Music", "Travel"]. You can use these attributes as filter criteria when you create segments. Attribute names are case sensitive.

An attribute name can contain up to 50 characters. An attribute value can contain up to 100 characters. When you define the name of a custom attribute, avoid using the following characters: number sign (#), colon (:), question mark (?), backslash (\), and slash (/). The Amazon Pinpoint console can't display attribute names that contain these characters. This restriction doesn't apply to attribute values.

$sel:endpointStatus:EndpointRequest', endpointRequest_endpointStatus - Specifies whether to send messages or push notifications to the endpoint. Valid values are: ACTIVE, messages are sent to the endpoint; and, INACTIVE, messages aren’t sent to the endpoint.

Amazon Pinpoint automatically sets this value to ACTIVE when you create an endpoint or update an existing endpoint. Amazon Pinpoint automatically sets this value to INACTIVE if you update another endpoint that has the same address specified by the Address property.

$sel:optOut:EndpointRequest', endpointRequest_optOut - Specifies whether the user who's associated with the endpoint has opted out of receiving messages and push notifications from you. Possible values are: ALL, the user has opted out and doesn't want to receive any messages or push notifications; and, NONE, the user hasn't opted out and wants to receive all messages and push notifications.

$sel:channelType:EndpointRequest', endpointRequest_channelType - The channel to use when sending messages or push notifications to the endpoint.

endpointRequest_requestId :: Lens' EndpointRequest (Maybe Text)

The unique identifier for the most recent request to update the endpoint.

endpointRequest_metrics :: Lens' EndpointRequest (Maybe (HashMap Text Double))

One or more custom metrics that your app reports to Amazon Pinpoint for the endpoint.

endpointRequest_location :: Lens' EndpointRequest (Maybe EndpointLocation)

The geographic information for the endpoint.

endpointRequest_demographic :: Lens' EndpointRequest (Maybe EndpointDemographic)

The demographic information for the endpoint, such as the time zone and platform.

endpointRequest_address :: Lens' EndpointRequest (Maybe Text)

The destination address for messages or push notifications that you send to the endpoint. The address varies by channel. For a push-notification channel, use the token provided by the push notification service, such as an Apple Push Notification service (APNs) device token or a Firebase Cloud Messaging (FCM) registration token. For the SMS channel, use a phone number in E.164 format, such as +12065550100. For the email channel, use an email address.

endpointRequest_effectiveDate :: Lens' EndpointRequest (Maybe Text)

The date and time, in ISO 8601 format, when the endpoint is updated.

endpointRequest_user :: Lens' EndpointRequest (Maybe EndpointUser)

One or more custom attributes that describe the user who's associated with the endpoint.

endpointRequest_attributes :: Lens' EndpointRequest (Maybe (HashMap Text [Text]))

One or more custom attributes that describe the endpoint by associating a name with an array of values. For example, the value of a custom attribute named Interests might be: ["Science", "Music", "Travel"]. You can use these attributes as filter criteria when you create segments. Attribute names are case sensitive.

An attribute name can contain up to 50 characters. An attribute value can contain up to 100 characters. When you define the name of a custom attribute, avoid using the following characters: number sign (#), colon (:), question mark (?), backslash (\), and slash (/). The Amazon Pinpoint console can't display attribute names that contain these characters. This restriction doesn't apply to attribute values.

endpointRequest_endpointStatus :: Lens' EndpointRequest (Maybe Text)

Specifies whether to send messages or push notifications to the endpoint. Valid values are: ACTIVE, messages are sent to the endpoint; and, INACTIVE, messages aren’t sent to the endpoint.

Amazon Pinpoint automatically sets this value to ACTIVE when you create an endpoint or update an existing endpoint. Amazon Pinpoint automatically sets this value to INACTIVE if you update another endpoint that has the same address specified by the Address property.

endpointRequest_optOut :: Lens' EndpointRequest (Maybe Text)

Specifies whether the user who's associated with the endpoint has opted out of receiving messages and push notifications from you. Possible values are: ALL, the user has opted out and doesn't want to receive any messages or push notifications; and, NONE, the user hasn't opted out and wants to receive all messages and push notifications.

endpointRequest_channelType :: Lens' EndpointRequest (Maybe ChannelType)

The channel to use when sending messages or push notifications to the endpoint.