Amazonka.Pinpoint.Types.EndpointResponse

Description

Documentation

data EndpointResponse

Provides information about the channel type and other settings for an endpoint.

See: newEndpointResponse smart constructor.

Constructors

EndpointResponse'

Fields

requestId :: Maybe Text

The unique identifier for the most recent request to update the endpoint.

metrics :: Maybe (HashMap Text Double)

One or more custom metrics that your app reports to Amazon Pinpoint for the endpoint.

location :: Maybe EndpointLocation

The geographic information for the endpoint.

demographic :: Maybe EndpointDemographic

The demographic information for the endpoint, such as the time zone and platform.

cohortId :: Maybe Text

A number from 0-99 that represents the cohort that the endpoint is assigned to. Endpoints are grouped into cohorts randomly, and each cohort contains approximately 1 percent of the endpoints for an application. Amazon Pinpoint assigns cohorts to the holdout or treatment allocations for campaigns.

address :: Maybe Text

The destination address for messages or push notifications that you send to the endpoint. The address varies by channel. For example, the address for a push-notification channel is typically the token provided by a push notification service, such as an Apple Push Notification service (APNs) device token or a Firebase Cloud Messaging (FCM) registration token. The address for the SMS channel is a phone number in E.164 format, such as +12065550100. The address for the email channel is an email address.

effectiveDate :: Maybe Text

The date and time, in ISO 8601 format, when the endpoint was last updated.

user :: Maybe EndpointUser

One or more custom user attributes that your app reports to Amazon Pinpoint for the user who's associated with the endpoint.

applicationId :: Maybe Text

The unique identifier for the application that's associated with the endpoint.

attributes :: Maybe (HashMap Text [Text])

One or more custom attributes that describe the endpoint by associating a name with an array of values. For example, the value of a custom attribute named Interests might be: ["Science", "Music", "Travel"]. You can use these attributes as filter criteria when you create segments.

endpointStatus :: Maybe Text

Specifies whether messages or push notifications are sent to the endpoint. Possible values are: ACTIVE, messages are sent to the endpoint; and, INACTIVE, messages aren’t sent to the endpoint.

Amazon Pinpoint automatically sets this value to ACTIVE when you create an endpoint or update an existing endpoint. Amazon Pinpoint automatically sets this value to INACTIVE if you update another endpoint that has the same address specified by the Address property.

optOut :: Maybe Text

Specifies whether the user who's associated with the endpoint has opted out of receiving messages and push notifications from you. Possible values are: ALL, the user has opted out and doesn't want to receive any messages or push notifications; and, NONE, the user hasn't opted out and wants to receive all messages and push notifications.

id :: Maybe Text

The unique identifier that you assigned to the endpoint. The identifier should be a globally unique identifier (GUID) to ensure that it doesn't conflict with other endpoint identifiers that are associated with the application.

creationDate :: Maybe Text

The date and time, in ISO 8601 format, when the endpoint was created.

channelType :: Maybe ChannelType

The channel that's used when sending messages or push notifications to the endpoint.

newEndpointResponse :: EndpointResponse

Create a value of EndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:EndpointResponse', endpointResponse_requestId - The unique identifier for the most recent request to update the endpoint.

$sel:metrics:EndpointResponse', endpointResponse_metrics - One or more custom metrics that your app reports to Amazon Pinpoint for the endpoint.

$sel:location:EndpointResponse', endpointResponse_location - The geographic information for the endpoint.

$sel:demographic:EndpointResponse', endpointResponse_demographic - The demographic information for the endpoint, such as the time zone and platform.

$sel:cohortId:EndpointResponse', endpointResponse_cohortId - A number from 0-99 that represents the cohort that the endpoint is assigned to. Endpoints are grouped into cohorts randomly, and each cohort contains approximately 1 percent of the endpoints for an application. Amazon Pinpoint assigns cohorts to the holdout or treatment allocations for campaigns.

$sel:address:EndpointResponse', endpointResponse_address - The destination address for messages or push notifications that you send to the endpoint. The address varies by channel. For example, the address for a push-notification channel is typically the token provided by a push notification service, such as an Apple Push Notification service (APNs) device token or a Firebase Cloud Messaging (FCM) registration token. The address for the SMS channel is a phone number in E.164 format, such as +12065550100. The address for the email channel is an email address.

$sel:effectiveDate:EndpointResponse', endpointResponse_effectiveDate - The date and time, in ISO 8601 format, when the endpoint was last updated.

$sel:user:EndpointResponse', endpointResponse_user - One or more custom user attributes that your app reports to Amazon Pinpoint for the user who's associated with the endpoint.

$sel:applicationId:EndpointResponse', endpointResponse_applicationId - The unique identifier for the application that's associated with the endpoint.

$sel:attributes:EndpointResponse', endpointResponse_attributes - One or more custom attributes that describe the endpoint by associating a name with an array of values. For example, the value of a custom attribute named Interests might be: ["Science", "Music", "Travel"]. You can use these attributes as filter criteria when you create segments.

$sel:endpointStatus:EndpointResponse', endpointResponse_endpointStatus - Specifies whether messages or push notifications are sent to the endpoint. Possible values are: ACTIVE, messages are sent to the endpoint; and, INACTIVE, messages aren’t sent to the endpoint.

Amazon Pinpoint automatically sets this value to ACTIVE when you create an endpoint or update an existing endpoint. Amazon Pinpoint automatically sets this value to INACTIVE if you update another endpoint that has the same address specified by the Address property.

$sel:optOut:EndpointResponse', endpointResponse_optOut - Specifies whether the user who's associated with the endpoint has opted out of receiving messages and push notifications from you. Possible values are: ALL, the user has opted out and doesn't want to receive any messages or push notifications; and, NONE, the user hasn't opted out and wants to receive all messages and push notifications.

$sel:id:EndpointResponse', endpointResponse_id - The unique identifier that you assigned to the endpoint. The identifier should be a globally unique identifier (GUID) to ensure that it doesn't conflict with other endpoint identifiers that are associated with the application.

$sel:creationDate:EndpointResponse', endpointResponse_creationDate - The date and time, in ISO 8601 format, when the endpoint was created.

$sel:channelType:EndpointResponse', endpointResponse_channelType - The channel that's used when sending messages or push notifications to the endpoint.

endpointResponse_requestId :: Lens' EndpointResponse (Maybe Text)

The unique identifier for the most recent request to update the endpoint.

endpointResponse_metrics :: Lens' EndpointResponse (Maybe (HashMap Text Double))

One or more custom metrics that your app reports to Amazon Pinpoint for the endpoint.

endpointResponse_location :: Lens' EndpointResponse (Maybe EndpointLocation)

The geographic information for the endpoint.

endpointResponse_demographic :: Lens' EndpointResponse (Maybe EndpointDemographic)

The demographic information for the endpoint, such as the time zone and platform.

endpointResponse_cohortId :: Lens' EndpointResponse (Maybe Text)

A number from 0-99 that represents the cohort that the endpoint is assigned to. Endpoints are grouped into cohorts randomly, and each cohort contains approximately 1 percent of the endpoints for an application. Amazon Pinpoint assigns cohorts to the holdout or treatment allocations for campaigns.

endpointResponse_address :: Lens' EndpointResponse (Maybe Text)

The destination address for messages or push notifications that you send to the endpoint. The address varies by channel. For example, the address for a push-notification channel is typically the token provided by a push notification service, such as an Apple Push Notification service (APNs) device token or a Firebase Cloud Messaging (FCM) registration token. The address for the SMS channel is a phone number in E.164 format, such as +12065550100. The address for the email channel is an email address.

endpointResponse_effectiveDate :: Lens' EndpointResponse (Maybe Text)

The date and time, in ISO 8601 format, when the endpoint was last updated.

endpointResponse_user :: Lens' EndpointResponse (Maybe EndpointUser)

One or more custom user attributes that your app reports to Amazon Pinpoint for the user who's associated with the endpoint.

endpointResponse_applicationId :: Lens' EndpointResponse (Maybe Text)

The unique identifier for the application that's associated with the endpoint.

endpointResponse_attributes :: Lens' EndpointResponse (Maybe (HashMap Text [Text]))

One or more custom attributes that describe the endpoint by associating a name with an array of values. For example, the value of a custom attribute named Interests might be: ["Science", "Music", "Travel"]. You can use these attributes as filter criteria when you create segments.

endpointResponse_endpointStatus :: Lens' EndpointResponse (Maybe Text)

Specifies whether messages or push notifications are sent to the endpoint. Possible values are: ACTIVE, messages are sent to the endpoint; and, INACTIVE, messages aren’t sent to the endpoint.

Amazon Pinpoint automatically sets this value to ACTIVE when you create an endpoint or update an existing endpoint. Amazon Pinpoint automatically sets this value to INACTIVE if you update another endpoint that has the same address specified by the Address property.

endpointResponse_optOut :: Lens' EndpointResponse (Maybe Text)

Specifies whether the user who's associated with the endpoint has opted out of receiving messages and push notifications from you. Possible values are: ALL, the user has opted out and doesn't want to receive any messages or push notifications; and, NONE, the user hasn't opted out and wants to receive all messages and push notifications.

endpointResponse_id :: Lens' EndpointResponse (Maybe Text)

The unique identifier that you assigned to the endpoint. The identifier should be a globally unique identifier (GUID) to ensure that it doesn't conflict with other endpoint identifiers that are associated with the application.

endpointResponse_creationDate :: Lens' EndpointResponse (Maybe Text)

The date and time, in ISO 8601 format, when the endpoint was created.

endpointResponse_channelType :: Lens' EndpointResponse (Maybe ChannelType)

The channel that's used when sending messages or push notifications to the endpoint.