Amazonka.Pinpoint.Types.JourneyResponse

Description

Documentation

data JourneyResponse

Provides information about the status, configuration, and other settings for a journey.

See: newJourneyResponse smart constructor.

Constructors

JourneyResponse'

Fields

state :: Maybe State

The current status of the journey. Possible values are:

DRAFT - The journey is being developed and hasn't been published yet.

ACTIVE - The journey has been developed and published. Depending on the journey's schedule, the journey may currently be running or scheduled to start running at a later time. If a journey's status is ACTIVE, you can't add, change, or remove activities from it.

COMPLETED - The journey has been published and has finished running. All participants have entered the journey and no participants are waiting to complete the journey or any activities in the journey.

CANCELLED - The journey has been stopped. If a journey's status is CANCELLED, you can't add, change, or remove activities or segment settings from the journey.

CLOSED - The journey has been published and has started running. It may have also passed its scheduled end time, or passed its scheduled start time and a refresh frequency hasn't been specified for it. If a journey's status is CLOSED, you can't add participants to it, and no existing participants can enter the journey for the first time. However, any existing participants who are currently waiting to start an activity may continue the journey.

lastModifiedDate :: Maybe Text

The date, in ISO 8601 format, when the journey was last modified.

schedule :: Maybe JourneySchedule

The schedule settings for the journey.

localTime :: Maybe Bool

Specifies whether the journey's scheduled start and end times use each participant's local time. If this value is true, the schedule uses each participant's local time.

activities :: Maybe (HashMap Text Activity)

A map that contains a set of Activity objects, one object for each activity in the journey. For each Activity object, the key is the unique identifier (string) for an activity and the value is the settings for the activity.

refreshOnSegmentUpdate :: Maybe Bool

Specifies whether a journey should be refreshed on segment update.

limits :: Maybe JourneyLimits

The messaging and entry limits for the journey.

waitForQuietTime :: Maybe Bool

Specifies whether endpoints in quiet hours should enter a wait till the end of their quiet hours.

quietTime :: Maybe QuietTime

The quiet time settings for the journey. Quiet time is a specific time range when a journey doesn't send messages to participants, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint for the participant is set to a valid value.

The current time in the participant's time zone is later than or equal to the time specified by the QuietTime.Start property for the journey.

The current time in the participant's time zone is earlier than or equal to the time specified by the QuietTime.End property for the journey.

If any of the preceding conditions isn't met, the participant will receive messages from the journey, even if quiet time is enabled.

startActivity :: Maybe Text

The unique identifier for the first activity in the journey.

creationDate :: Maybe Text

The date, in ISO 8601 format, when the journey was created.

startCondition :: Maybe StartCondition

The segment that defines which users are participants in the journey.

journeyChannelSettings :: Maybe JourneyChannelSettings

The channel-specific configurations for the journey.

refreshFrequency :: Maybe Text

The frequency with which Amazon Pinpoint evaluates segment and event data for the journey, as a duration in ISO 8601 format.

tags :: Maybe (HashMap Text Text)

This object is not used or supported.

name :: Text

The name of the journey.

id :: Text

The unique identifier for the journey.

applicationId :: Text

The unique identifier for the application that the journey applies to.

newJourneyResponse

Arguments

:: Text
$sel:name:JourneyResponse'

-> Text
$sel:id:JourneyResponse'

-> Text
$sel:applicationId:JourneyResponse'

-> JourneyResponse

Create a value of JourneyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:JourneyResponse', journeyResponse_state - The current status of the journey. Possible values are:

DRAFT - The journey is being developed and hasn't been published yet.

ACTIVE - The journey has been developed and published. Depending on the journey's schedule, the journey may currently be running or scheduled to start running at a later time. If a journey's status is ACTIVE, you can't add, change, or remove activities from it.

COMPLETED - The journey has been published and has finished running. All participants have entered the journey and no participants are waiting to complete the journey or any activities in the journey.

CANCELLED - The journey has been stopped. If a journey's status is CANCELLED, you can't add, change, or remove activities or segment settings from the journey.

CLOSED - The journey has been published and has started running. It may have also passed its scheduled end time, or passed its scheduled start time and a refresh frequency hasn't been specified for it. If a journey's status is CLOSED, you can't add participants to it, and no existing participants can enter the journey for the first time. However, any existing participants who are currently waiting to start an activity may continue the journey.

$sel:lastModifiedDate:JourneyResponse', journeyResponse_lastModifiedDate - The date, in ISO 8601 format, when the journey was last modified.

$sel:schedule:JourneyResponse', journeyResponse_schedule - The schedule settings for the journey.

$sel:localTime:JourneyResponse', journeyResponse_localTime - Specifies whether the journey's scheduled start and end times use each participant's local time. If this value is true, the schedule uses each participant's local time.

$sel:activities:JourneyResponse', journeyResponse_activities - A map that contains a set of Activity objects, one object for each activity in the journey. For each Activity object, the key is the unique identifier (string) for an activity and the value is the settings for the activity.

$sel:refreshOnSegmentUpdate:JourneyResponse', journeyResponse_refreshOnSegmentUpdate - Specifies whether a journey should be refreshed on segment update.

$sel:limits:JourneyResponse', journeyResponse_limits - The messaging and entry limits for the journey.

$sel:waitForQuietTime:JourneyResponse', journeyResponse_waitForQuietTime - Specifies whether endpoints in quiet hours should enter a wait till the end of their quiet hours.

$sel:quietTime:JourneyResponse', journeyResponse_quietTime - The quiet time settings for the journey. Quiet time is a specific time range when a journey doesn't send messages to participants, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint for the participant is set to a valid value.

The current time in the participant's time zone is later than or equal to the time specified by the QuietTime.Start property for the journey.

The current time in the participant's time zone is earlier than or equal to the time specified by the QuietTime.End property for the journey.

If any of the preceding conditions isn't met, the participant will receive messages from the journey, even if quiet time is enabled.

$sel:startActivity:JourneyResponse', journeyResponse_startActivity - The unique identifier for the first activity in the journey.

$sel:creationDate:JourneyResponse', journeyResponse_creationDate - The date, in ISO 8601 format, when the journey was created.

$sel:startCondition:JourneyResponse', journeyResponse_startCondition - The segment that defines which users are participants in the journey.

$sel:journeyChannelSettings:JourneyResponse', journeyResponse_journeyChannelSettings - The channel-specific configurations for the journey.

$sel:refreshFrequency:JourneyResponse', journeyResponse_refreshFrequency - The frequency with which Amazon Pinpoint evaluates segment and event data for the journey, as a duration in ISO 8601 format.

$sel:tags:JourneyResponse', journeyResponse_tags - This object is not used or supported.

$sel:name:JourneyResponse', journeyResponse_name - The name of the journey.

$sel:id:JourneyResponse', journeyResponse_id - The unique identifier for the journey.

$sel:applicationId:JourneyResponse', journeyResponse_applicationId - The unique identifier for the application that the journey applies to.

journeyResponse_state :: Lens' JourneyResponse (Maybe State)

The current status of the journey. Possible values are:

DRAFT - The journey is being developed and hasn't been published yet.

ACTIVE - The journey has been developed and published. Depending on the journey's schedule, the journey may currently be running or scheduled to start running at a later time. If a journey's status is ACTIVE, you can't add, change, or remove activities from it.

COMPLETED - The journey has been published and has finished running. All participants have entered the journey and no participants are waiting to complete the journey or any activities in the journey.

CANCELLED - The journey has been stopped. If a journey's status is CANCELLED, you can't add, change, or remove activities or segment settings from the journey.

CLOSED - The journey has been published and has started running. It may have also passed its scheduled end time, or passed its scheduled start time and a refresh frequency hasn't been specified for it. If a journey's status is CLOSED, you can't add participants to it, and no existing participants can enter the journey for the first time. However, any existing participants who are currently waiting to start an activity may continue the journey.

journeyResponse_lastModifiedDate :: Lens' JourneyResponse (Maybe Text)

The date, in ISO 8601 format, when the journey was last modified.

journeyResponse_schedule :: Lens' JourneyResponse (Maybe JourneySchedule)

The schedule settings for the journey.

journeyResponse_localTime :: Lens' JourneyResponse (Maybe Bool)

Specifies whether the journey's scheduled start and end times use each participant's local time. If this value is true, the schedule uses each participant's local time.

journeyResponse_activities :: Lens' JourneyResponse (Maybe (HashMap Text Activity))

A map that contains a set of Activity objects, one object for each activity in the journey. For each Activity object, the key is the unique identifier (string) for an activity and the value is the settings for the activity.

journeyResponse_refreshOnSegmentUpdate :: Lens' JourneyResponse (Maybe Bool)

Specifies whether a journey should be refreshed on segment update.

journeyResponse_limits :: Lens' JourneyResponse (Maybe JourneyLimits)

The messaging and entry limits for the journey.

journeyResponse_waitForQuietTime :: Lens' JourneyResponse (Maybe Bool)

Specifies whether endpoints in quiet hours should enter a wait till the end of their quiet hours.

journeyResponse_quietTime :: Lens' JourneyResponse (Maybe QuietTime)

The quiet time settings for the journey. Quiet time is a specific time range when a journey doesn't send messages to participants, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint for the participant is set to a valid value.

The current time in the participant's time zone is later than or equal to the time specified by the QuietTime.Start property for the journey.

The current time in the participant's time zone is earlier than or equal to the time specified by the QuietTime.End property for the journey.

If any of the preceding conditions isn't met, the participant will receive messages from the journey, even if quiet time is enabled.

journeyResponse_startActivity :: Lens' JourneyResponse (Maybe Text)

The unique identifier for the first activity in the journey.

journeyResponse_creationDate :: Lens' JourneyResponse (Maybe Text)

The date, in ISO 8601 format, when the journey was created.

journeyResponse_startCondition :: Lens' JourneyResponse (Maybe StartCondition)

The segment that defines which users are participants in the journey.

journeyResponse_journeyChannelSettings :: Lens' JourneyResponse (Maybe JourneyChannelSettings)

The channel-specific configurations for the journey.

journeyResponse_refreshFrequency :: Lens' JourneyResponse (Maybe Text)

The frequency with which Amazon Pinpoint evaluates segment and event data for the journey, as a duration in ISO 8601 format.

journeyResponse_tags :: Lens' JourneyResponse (Maybe (HashMap Text Text))

This object is not used or supported.

journeyResponse_name :: Lens' JourneyResponse Text

The name of the journey.

journeyResponse_id :: Lens' JourneyResponse Text

The unique identifier for the journey.

journeyResponse_applicationId :: Lens' JourneyResponse Text

The unique identifier for the application that the journey applies to.