Amazonka.Pinpoint.Types.JourneySchedule

Description

Documentation

data JourneySchedule

Specifies the schedule settings for a journey.

See: newJourneySchedule smart constructor.

Constructors

JourneySchedule'

Fields

startTime :: Maybe POSIX

The scheduled time, in ISO 8601 format, when the journey began or will begin.

endTime :: Maybe POSIX

The scheduled time, in ISO 8601 format, when the journey ended or will end.

timezone :: Maybe Text

The starting UTC offset for the journey schedule, if the value of the journey's LocalTime property is true. Valid values are: UTC, UTC+01, UTC+02, UTC+03, UTC+03:30, UTC+04, UTC+04:30, UTC+05, UTC+05:30, UTC+05:45, UTC+06, UTC+06:30, UTC+07, UTC+08, UTC+08:45, UTC+09, UTC+09:30, UTC+10, UTC+10:30, UTC+11, UTC+12, UTC+12:45, UTC+13, UTC+13:45, UTC-02, UTC-02:30, UTC-03, UTC-03:30, UTC-04, UTC-05, UTC-06, UTC-07, UTC-08, UTC-09, UTC-09:30, UTC-10, and UTC-11.

newJourneySchedule :: JourneySchedule

Create a value of JourneySchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:JourneySchedule', journeySchedule_startTime - The scheduled time, in ISO 8601 format, when the journey began or will begin.

$sel:endTime:JourneySchedule', journeySchedule_endTime - The scheduled time, in ISO 8601 format, when the journey ended or will end.

$sel:timezone:JourneySchedule', journeySchedule_timezone - The starting UTC offset for the journey schedule, if the value of the journey's LocalTime property is true. Valid values are: UTC, UTC+01, UTC+02, UTC+03, UTC+03:30, UTC+04, UTC+04:30, UTC+05, UTC+05:30, UTC+05:45, UTC+06, UTC+06:30, UTC+07, UTC+08, UTC+08:45, UTC+09, UTC+09:30, UTC+10, UTC+10:30, UTC+11, UTC+12, UTC+12:45, UTC+13, UTC+13:45, UTC-02, UTC-02:30, UTC-03, UTC-03:30, UTC-04, UTC-05, UTC-06, UTC-07, UTC-08, UTC-09, UTC-09:30, UTC-10, and UTC-11.

journeySchedule_startTime :: Lens' JourneySchedule (Maybe UTCTime)

The scheduled time, in ISO 8601 format, when the journey began or will begin.

journeySchedule_endTime :: Lens' JourneySchedule (Maybe UTCTime)

The scheduled time, in ISO 8601 format, when the journey ended or will end.

journeySchedule_timezone :: Lens' JourneySchedule (Maybe Text)

The starting UTC offset for the journey schedule, if the value of the journey's LocalTime property is true. Valid values are: UTC, UTC+01, UTC+02, UTC+03, UTC+03:30, UTC+04, UTC+04:30, UTC+05, UTC+05:30, UTC+05:45, UTC+06, UTC+06:30, UTC+07, UTC+08, UTC+08:45, UTC+09, UTC+09:30, UTC+10, UTC+10:30, UTC+11, UTC+12, UTC+12:45, UTC+13, UTC+13:45, UTC-02, UTC-02:30, UTC-03, UTC-03:30, UTC-04, UTC-05, UTC-06, UTC-07, UTC-08, UTC-09, UTC-09:30, UTC-10, and UTC-11.