Amazonka.Pinpoint.Types.Message

Description

Documentation

data Message

Specifies the content and settings for a push notification that's sent to recipients of a campaign.

See: newMessage smart constructor.

Constructors

Message'

Fields

silentPush :: Maybe Bool

Specifies whether the notification is a silent push notification, which is a push notification that doesn't display on a recipient's device. Silent push notifications can be used for cases such as updating an app's configuration, displaying messages in an in-app message center, or supporting phone home functionality.

imageIconUrl :: Maybe Text

The URL of the image to display as the push-notification icon, such as the icon for the app.

rawContent :: Maybe Text

The raw, JSON-formatted string to use as the payload for the notification message. If specified, this value overrides all other content for the message.

body :: Maybe Text

The body of the notification message. The maximum number of characters is 200.

timeToLive :: Maybe Int

The number of seconds that the push-notification service should keep the message, if the service is unable to deliver the notification the first time. This value is converted to an expiration value when it's sent to a push-notification service. If this value is 0, the service treats the notification as if it expires immediately and the service doesn't store or try to deliver the notification again.

This value doesn't apply to messages that are sent through the Amazon Device Messaging (ADM) service.

imageSmallIconUrl :: Maybe Text

The URL of the image to display as the small, push-notification icon, such as a small version of the icon for the app.

jsonBody :: Maybe Text

The JSON payload to use for a silent push notification.

url :: Maybe Text

The URL to open in a recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.

action :: Maybe Action

The action to occur if a recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of iOS and Android.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

imageUrl :: Maybe Text

The URL of an image to display in the push notification.

mediaUrl :: Maybe Text

The URL of the image or video to display in the push notification.

title :: Maybe Text

The title to display above the notification message on a recipient's device.

newMessage :: Message

Create a value of Message with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:silentPush:Message', message_silentPush - Specifies whether the notification is a silent push notification, which is a push notification that doesn't display on a recipient's device. Silent push notifications can be used for cases such as updating an app's configuration, displaying messages in an in-app message center, or supporting phone home functionality.

$sel:imageIconUrl:Message', message_imageIconUrl - The URL of the image to display as the push-notification icon, such as the icon for the app.

$sel:rawContent:Message', message_rawContent - The raw, JSON-formatted string to use as the payload for the notification message. If specified, this value overrides all other content for the message.

$sel:body:Message', message_body - The body of the notification message. The maximum number of characters is 200.

$sel:timeToLive:Message', message_timeToLive - The number of seconds that the push-notification service should keep the message, if the service is unable to deliver the notification the first time. This value is converted to an expiration value when it's sent to a push-notification service. If this value is 0, the service treats the notification as if it expires immediately and the service doesn't store or try to deliver the notification again.

This value doesn't apply to messages that are sent through the Amazon Device Messaging (ADM) service.

$sel:imageSmallIconUrl:Message', message_imageSmallIconUrl - The URL of the image to display as the small, push-notification icon, such as a small version of the icon for the app.

$sel:jsonBody:Message', message_jsonBody - The JSON payload to use for a silent push notification.

$sel:url:Message', message_url - The URL to open in a recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.

$sel:action:Message', message_action - The action to occur if a recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of iOS and Android.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

$sel:imageUrl:Message', message_imageUrl - The URL of an image to display in the push notification.

$sel:mediaUrl:Message', message_mediaUrl - The URL of the image or video to display in the push notification.

$sel:title:Message', message_title - The title to display above the notification message on a recipient's device.

message_silentPush :: Lens' Message (Maybe Bool)

Specifies whether the notification is a silent push notification, which is a push notification that doesn't display on a recipient's device. Silent push notifications can be used for cases such as updating an app's configuration, displaying messages in an in-app message center, or supporting phone home functionality.

message_imageIconUrl :: Lens' Message (Maybe Text)

The URL of the image to display as the push-notification icon, such as the icon for the app.

message_rawContent :: Lens' Message (Maybe Text)

The raw, JSON-formatted string to use as the payload for the notification message. If specified, this value overrides all other content for the message.

message_body :: Lens' Message (Maybe Text)

The body of the notification message. The maximum number of characters is 200.

message_timeToLive :: Lens' Message (Maybe Int)

The number of seconds that the push-notification service should keep the message, if the service is unable to deliver the notification the first time. This value is converted to an expiration value when it's sent to a push-notification service. If this value is 0, the service treats the notification as if it expires immediately and the service doesn't store or try to deliver the notification again.

This value doesn't apply to messages that are sent through the Amazon Device Messaging (ADM) service.

message_imageSmallIconUrl :: Lens' Message (Maybe Text)

The URL of the image to display as the small, push-notification icon, such as a small version of the icon for the app.

message_jsonBody :: Lens' Message (Maybe Text)

The JSON payload to use for a silent push notification.

message_url :: Lens' Message (Maybe Text)

The URL to open in a recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.

message_action :: Lens' Message (Maybe Action)

The action to occur if a recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of iOS and Android.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

message_imageUrl :: Lens' Message (Maybe Text)

The URL of an image to display in the push notification.

message_mediaUrl :: Lens' Message (Maybe Text)

The URL of the image or video to display in the push notification.

message_title :: Lens' Message (Maybe Text)

The title to display above the notification message on a recipient's device.