Amazonka.Pinpoint.Types.MessageResult

Description

Documentation

data MessageResult

Provides information about the results of sending a message directly to an endpoint address.

See: newMessageResult smart constructor.

Constructors

MessageResult'

Fields

statusMessage :: Maybe Text

The status message for delivering the message.

updatedToken :: Maybe Text

For push notifications that are sent through the GCM channel, specifies whether the endpoint's device registration token was updated as part of delivering the message.

messageId :: Maybe Text

The unique identifier for the message that was sent.

deliveryStatus :: DeliveryStatus

The delivery status of the message. Possible values are:

DUPLICATE - The endpoint address is a duplicate of another endpoint address. Amazon Pinpoint won't attempt to send the message again.

OPT_OUT - The user who's associated with the endpoint address has opted out of receiving messages from you. Amazon Pinpoint won't attempt to send the message again.

PERMANENT_FAILURE - An error occurred when delivering the message to the endpoint address. Amazon Pinpoint won't attempt to send the message again.

SUCCESSFUL - The message was successfully delivered to the endpoint address.

TEMPORARY_FAILURE - A temporary error occurred. Amazon Pinpoint won't attempt to send the message again.

THROTTLED - Amazon Pinpoint throttled the operation to send the message to the endpoint address.

TIMEOUT - The message couldn't be sent within the timeout period.

UNKNOWN_FAILURE - An unknown error occurred.

statusCode :: Int

The downstream service status code for delivering the message.

newMessageResult

Arguments

:: DeliveryStatus
$sel:deliveryStatus:MessageResult'

-> Int
$sel:statusCode:MessageResult'

-> MessageResult

Create a value of MessageResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:MessageResult', messageResult_statusMessage - The status message for delivering the message.

$sel:updatedToken:MessageResult', messageResult_updatedToken - For push notifications that are sent through the GCM channel, specifies whether the endpoint's device registration token was updated as part of delivering the message.

$sel:messageId:MessageResult', messageResult_messageId - The unique identifier for the message that was sent.

$sel:deliveryStatus:MessageResult', messageResult_deliveryStatus - The delivery status of the message. Possible values are:

DUPLICATE - The endpoint address is a duplicate of another endpoint address. Amazon Pinpoint won't attempt to send the message again.

OPT_OUT - The user who's associated with the endpoint address has opted out of receiving messages from you. Amazon Pinpoint won't attempt to send the message again.

PERMANENT_FAILURE - An error occurred when delivering the message to the endpoint address. Amazon Pinpoint won't attempt to send the message again.

SUCCESSFUL - The message was successfully delivered to the endpoint address.

TEMPORARY_FAILURE - A temporary error occurred. Amazon Pinpoint won't attempt to send the message again.

THROTTLED - Amazon Pinpoint throttled the operation to send the message to the endpoint address.

TIMEOUT - The message couldn't be sent within the timeout period.

UNKNOWN_FAILURE - An unknown error occurred.

$sel:statusCode:MessageResult', messageResult_statusCode - The downstream service status code for delivering the message.

messageResult_statusMessage :: Lens' MessageResult (Maybe Text)

The status message for delivering the message.

messageResult_updatedToken :: Lens' MessageResult (Maybe Text)

For push notifications that are sent through the GCM channel, specifies whether the endpoint's device registration token was updated as part of delivering the message.

messageResult_messageId :: Lens' MessageResult (Maybe Text)

The unique identifier for the message that was sent.

messageResult_deliveryStatus :: Lens' MessageResult DeliveryStatus

The delivery status of the message. Possible values are:

DUPLICATE - The endpoint address is a duplicate of another endpoint address. Amazon Pinpoint won't attempt to send the message again.

OPT_OUT - The user who's associated with the endpoint address has opted out of receiving messages from you. Amazon Pinpoint won't attempt to send the message again.

PERMANENT_FAILURE - An error occurred when delivering the message to the endpoint address. Amazon Pinpoint won't attempt to send the message again.

SUCCESSFUL - The message was successfully delivered to the endpoint address.

TEMPORARY_FAILURE - A temporary error occurred. Amazon Pinpoint won't attempt to send the message again.

THROTTLED - Amazon Pinpoint throttled the operation to send the message to the endpoint address.

TIMEOUT - The message couldn't be sent within the timeout period.

UNKNOWN_FAILURE - An unknown error occurred.

messageResult_statusCode :: Lens' MessageResult Int

The downstream service status code for delivering the message.