Amazonka.Pinpoint.Types.PushNotificationTemplateResponse

Description

Documentation

data PushNotificationTemplateResponse

Provides information about the content and settings for a message template that can be used in messages that are sent through a push notification channel.

See: newPushNotificationTemplateResponse smart constructor.

Constructors

PushNotificationTemplateResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the message template.

default' :: Maybe DefaultPushNotificationTemplate

The default message template that's used for push notification channels.

templateDescription :: Maybe Text

The custom description of the message template.

gcm :: Maybe AndroidPushNotificationTemplate

The message template that's used for the GCM channel, which is used to send notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

apns :: Maybe APNSPushNotificationTemplate

The message template that's used for the APNs (Apple Push Notification service) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

defaultSubstitutions :: Maybe Text

The JSON object that specifies the default values that are used for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

version :: Maybe Text

The unique identifier, as an integer, for the active version of the message template, or the version of the template that you specified by using the version parameter in your request.

adm :: Maybe AndroidPushNotificationTemplate

The message template that's used for the ADM (Amazon Device Messaging) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

baidu :: Maybe AndroidPushNotificationTemplate

The message template that's used for the Baidu (Baidu Cloud Push) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

recommenderId :: Maybe Text

The unique identifier for the recommender model that's used by the message template.

tags :: Maybe (HashMap Text Text)

A string-to-string map of key-value pairs that identifies the tags that are associated with the message template. Each tag consists of a required tag key and an associated tag value.

lastModifiedDate :: Text

The date, in ISO 8601 format, when the message template was last modified.

creationDate :: Text

The date, in ISO 8601 format, when the message template was created.

templateType :: TemplateType

The type of channel that the message template is designed for. For a push notification template, this value is PUSH.

templateName :: Text

The name of the message template.

newPushNotificationTemplateResponse

Arguments

:: Text
$sel:lastModifiedDate:PushNotificationTemplateResponse'

-> Text
$sel:creationDate:PushNotificationTemplateResponse'

-> TemplateType
$sel:templateType:PushNotificationTemplateResponse'

-> Text
$sel:templateName:PushNotificationTemplateResponse'

-> PushNotificationTemplateResponse

Create a value of PushNotificationTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:PushNotificationTemplateResponse', pushNotificationTemplateResponse_arn - The Amazon Resource Name (ARN) of the message template.

$sel:default':PushNotificationTemplateResponse', pushNotificationTemplateResponse_default - The default message template that's used for push notification channels.

$sel:templateDescription:PushNotificationTemplateResponse', pushNotificationTemplateResponse_templateDescription - The custom description of the message template.

$sel:gcm:PushNotificationTemplateResponse', pushNotificationTemplateResponse_gcm - The message template that's used for the GCM channel, which is used to send notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

$sel:apns:PushNotificationTemplateResponse', pushNotificationTemplateResponse_apns - The message template that's used for the APNs (Apple Push Notification service) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

$sel:defaultSubstitutions:PushNotificationTemplateResponse', pushNotificationTemplateResponse_defaultSubstitutions - The JSON object that specifies the default values that are used for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

$sel:version:PushNotificationTemplateResponse', pushNotificationTemplateResponse_version - The unique identifier, as an integer, for the active version of the message template, or the version of the template that you specified by using the version parameter in your request.

$sel:adm:PushNotificationTemplateResponse', pushNotificationTemplateResponse_adm - The message template that's used for the ADM (Amazon Device Messaging) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

$sel:baidu:PushNotificationTemplateResponse', pushNotificationTemplateResponse_baidu - The message template that's used for the Baidu (Baidu Cloud Push) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

$sel:recommenderId:PushNotificationTemplateResponse', pushNotificationTemplateResponse_recommenderId - The unique identifier for the recommender model that's used by the message template.

$sel:tags:PushNotificationTemplateResponse', pushNotificationTemplateResponse_tags - A string-to-string map of key-value pairs that identifies the tags that are associated with the message template. Each tag consists of a required tag key and an associated tag value.

$sel:lastModifiedDate:PushNotificationTemplateResponse', pushNotificationTemplateResponse_lastModifiedDate - The date, in ISO 8601 format, when the message template was last modified.

$sel:creationDate:PushNotificationTemplateResponse', pushNotificationTemplateResponse_creationDate - The date, in ISO 8601 format, when the message template was created.

$sel:templateType:PushNotificationTemplateResponse', pushNotificationTemplateResponse_templateType - The type of channel that the message template is designed for. For a push notification template, this value is PUSH.

$sel:templateName:PushNotificationTemplateResponse', pushNotificationTemplateResponse_templateName - The name of the message template.

pushNotificationTemplateResponse_arn :: Lens' PushNotificationTemplateResponse (Maybe Text)

The Amazon Resource Name (ARN) of the message template.

pushNotificationTemplateResponse_default :: Lens' PushNotificationTemplateResponse (Maybe DefaultPushNotificationTemplate)

The default message template that's used for push notification channels.

pushNotificationTemplateResponse_templateDescription :: Lens' PushNotificationTemplateResponse (Maybe Text)

The custom description of the message template.

pushNotificationTemplateResponse_gcm :: Lens' PushNotificationTemplateResponse (Maybe AndroidPushNotificationTemplate)

The message template that's used for the GCM channel, which is used to send notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

pushNotificationTemplateResponse_apns :: Lens' PushNotificationTemplateResponse (Maybe APNSPushNotificationTemplate)

The message template that's used for the APNs (Apple Push Notification service) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

pushNotificationTemplateResponse_defaultSubstitutions :: Lens' PushNotificationTemplateResponse (Maybe Text)

The JSON object that specifies the default values that are used for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

pushNotificationTemplateResponse_version :: Lens' PushNotificationTemplateResponse (Maybe Text)

The unique identifier, as an integer, for the active version of the message template, or the version of the template that you specified by using the version parameter in your request.

pushNotificationTemplateResponse_adm :: Lens' PushNotificationTemplateResponse (Maybe AndroidPushNotificationTemplate)

The message template that's used for the ADM (Amazon Device Messaging) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

pushNotificationTemplateResponse_baidu :: Lens' PushNotificationTemplateResponse (Maybe AndroidPushNotificationTemplate)

The message template that's used for the Baidu (Baidu Cloud Push) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

pushNotificationTemplateResponse_recommenderId :: Lens' PushNotificationTemplateResponse (Maybe Text)

The unique identifier for the recommender model that's used by the message template.

pushNotificationTemplateResponse_tags :: Lens' PushNotificationTemplateResponse (Maybe (HashMap Text Text))

A string-to-string map of key-value pairs that identifies the tags that are associated with the message template. Each tag consists of a required tag key and an associated tag value.

pushNotificationTemplateResponse_lastModifiedDate :: Lens' PushNotificationTemplateResponse Text

The date, in ISO 8601 format, when the message template was last modified.

pushNotificationTemplateResponse_creationDate :: Lens' PushNotificationTemplateResponse Text

The date, in ISO 8601 format, when the message template was created.

pushNotificationTemplateResponse_templateType :: Lens' PushNotificationTemplateResponse TemplateType

The type of channel that the message template is designed for. For a push notification template, this value is PUSH.

pushNotificationTemplateResponse_templateName :: Lens' PushNotificationTemplateResponse Text

The name of the message template.