Amazonka.Pinpoint.Types.SMSChannelResponse

Description

Documentation

data SMSChannelResponse

Provides information about the status and settings of the SMS channel for an application.

See: newSMSChannelResponse smart constructor.

Constructors

SMSChannelResponse'

Fields

shortCode :: Maybe Text

The registered short code to use when you send messages through the SMS channel.

lastModifiedDate :: Maybe Text

The date and time, in ISO 8601 format, when the SMS channel was last modified.

enabled :: Maybe Bool

Specifies whether the SMS channel is enabled for the application.

senderId :: Maybe Text

The identity that displays on recipients' devices when they receive messages from the SMS channel.

transactionalMessagesPerSecond :: Maybe Int

The maximum number of transactional messages that you can send through the SMS channel each second.

promotionalMessagesPerSecond :: Maybe Int

The maximum number of promotional messages that you can send through the SMS channel each second.

isArchived :: Maybe Bool

Specifies whether the SMS channel is archived.

applicationId :: Maybe Text

The unique identifier for the application that the SMS channel applies to.

version :: Maybe Int

The current version of the SMS channel.

id :: Maybe Text

(Deprecated) An identifier for the SMS channel. This property is retained only for backward compatibility.

creationDate :: Maybe Text

The date and time, in ISO 8601 format, when the SMS channel was enabled.

lastModifiedBy :: Maybe Text

The user who last modified the SMS channel.

hasCredential :: Maybe Bool

(Not used) This property is retained only for backward compatibility.

platform :: Text

The type of messaging or notification platform for the channel. For the SMS channel, this value is SMS.

newSMSChannelResponse

Arguments

:: Text
$sel:platform:SMSChannelResponse'

-> SMSChannelResponse

Create a value of SMSChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shortCode:SMSChannelResponse', sMSChannelResponse_shortCode - The registered short code to use when you send messages through the SMS channel.

$sel:lastModifiedDate:SMSChannelResponse', sMSChannelResponse_lastModifiedDate - The date and time, in ISO 8601 format, when the SMS channel was last modified.

$sel:enabled:SMSChannelResponse', sMSChannelResponse_enabled - Specifies whether the SMS channel is enabled for the application.

$sel:senderId:SMSChannelResponse', sMSChannelResponse_senderId - The identity that displays on recipients' devices when they receive messages from the SMS channel.

$sel:transactionalMessagesPerSecond:SMSChannelResponse', sMSChannelResponse_transactionalMessagesPerSecond - The maximum number of transactional messages that you can send through the SMS channel each second.

$sel:promotionalMessagesPerSecond:SMSChannelResponse', sMSChannelResponse_promotionalMessagesPerSecond - The maximum number of promotional messages that you can send through the SMS channel each second.

$sel:isArchived:SMSChannelResponse', sMSChannelResponse_isArchived - Specifies whether the SMS channel is archived.

$sel:applicationId:SMSChannelResponse', sMSChannelResponse_applicationId - The unique identifier for the application that the SMS channel applies to.

$sel:version:SMSChannelResponse', sMSChannelResponse_version - The current version of the SMS channel.

$sel:id:SMSChannelResponse', sMSChannelResponse_id - (Deprecated) An identifier for the SMS channel. This property is retained only for backward compatibility.

$sel:creationDate:SMSChannelResponse', sMSChannelResponse_creationDate - The date and time, in ISO 8601 format, when the SMS channel was enabled.

$sel:lastModifiedBy:SMSChannelResponse', sMSChannelResponse_lastModifiedBy - The user who last modified the SMS channel.

$sel:hasCredential:SMSChannelResponse', sMSChannelResponse_hasCredential - (Not used) This property is retained only for backward compatibility.

$sel:platform:SMSChannelResponse', sMSChannelResponse_platform - The type of messaging or notification platform for the channel. For the SMS channel, this value is SMS.

sMSChannelResponse_shortCode :: Lens' SMSChannelResponse (Maybe Text)

The registered short code to use when you send messages through the SMS channel.

sMSChannelResponse_lastModifiedDate :: Lens' SMSChannelResponse (Maybe Text)

The date and time, in ISO 8601 format, when the SMS channel was last modified.

sMSChannelResponse_enabled :: Lens' SMSChannelResponse (Maybe Bool)

Specifies whether the SMS channel is enabled for the application.

sMSChannelResponse_senderId :: Lens' SMSChannelResponse (Maybe Text)

The identity that displays on recipients' devices when they receive messages from the SMS channel.

sMSChannelResponse_transactionalMessagesPerSecond :: Lens' SMSChannelResponse (Maybe Int)

The maximum number of transactional messages that you can send through the SMS channel each second.

sMSChannelResponse_promotionalMessagesPerSecond :: Lens' SMSChannelResponse (Maybe Int)

The maximum number of promotional messages that you can send through the SMS channel each second.

sMSChannelResponse_isArchived :: Lens' SMSChannelResponse (Maybe Bool)

Specifies whether the SMS channel is archived.

sMSChannelResponse_applicationId :: Lens' SMSChannelResponse (Maybe Text)

The unique identifier for the application that the SMS channel applies to.

sMSChannelResponse_version :: Lens' SMSChannelResponse (Maybe Int)

The current version of the SMS channel.

sMSChannelResponse_id :: Lens' SMSChannelResponse (Maybe Text)

(Deprecated) An identifier for the SMS channel. This property is retained only for backward compatibility.

sMSChannelResponse_creationDate :: Lens' SMSChannelResponse (Maybe Text)

The date and time, in ISO 8601 format, when the SMS channel was enabled.

sMSChannelResponse_lastModifiedBy :: Lens' SMSChannelResponse (Maybe Text)

The user who last modified the SMS channel.

sMSChannelResponse_hasCredential :: Lens' SMSChannelResponse (Maybe Bool)

(Not used) This property is retained only for backward compatibility.

sMSChannelResponse_platform :: Lens' SMSChannelResponse Text

The type of messaging or notification platform for the channel. For the SMS channel, this value is SMS.