Amazonka.Pinpoint.Types.SegmentDemographics

Description

Documentation

data SegmentDemographics

Specifies demographic-based dimension settings for including or excluding endpoints from a segment. These settings derive from characteristics of endpoint devices, such as platform, make, and model.

See: newSegmentDemographics smart constructor.

Constructors

SegmentDemographics'

Fields

platform :: Maybe SetDimension

The device platform criteria for the segment.

appVersion :: Maybe SetDimension

The app version criteria for the segment.

channel :: Maybe SetDimension

The channel criteria for the segment.

model :: Maybe SetDimension

The device model criteria for the segment.

make :: Maybe SetDimension

The device make criteria for the segment.

deviceType :: Maybe SetDimension

The device type criteria for the segment.

newSegmentDemographics :: SegmentDemographics

Create a value of SegmentDemographics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:SegmentDemographics', segmentDemographics_platform - The device platform criteria for the segment.

$sel:appVersion:SegmentDemographics', segmentDemographics_appVersion - The app version criteria for the segment.

$sel:channel:SegmentDemographics', segmentDemographics_channel - The channel criteria for the segment.

$sel:model:SegmentDemographics', segmentDemographics_model - The device model criteria for the segment.

$sel:make:SegmentDemographics', segmentDemographics_make - The device make criteria for the segment.

$sel:deviceType:SegmentDemographics', segmentDemographics_deviceType - The device type criteria for the segment.

segmentDemographics_platform :: Lens' SegmentDemographics (Maybe SetDimension)

The device platform criteria for the segment.

segmentDemographics_appVersion :: Lens' SegmentDemographics (Maybe SetDimension)

The app version criteria for the segment.

segmentDemographics_channel :: Lens' SegmentDemographics (Maybe SetDimension)

The channel criteria for the segment.

segmentDemographics_model :: Lens' SegmentDemographics (Maybe SetDimension)

The device model criteria for the segment.

segmentDemographics_make :: Lens' SegmentDemographics (Maybe SetDimension)

The device make criteria for the segment.

segmentDemographics_deviceType :: Lens' SegmentDemographics (Maybe SetDimension)

The device type criteria for the segment.