Amazonka.Pinpoint.Types.SegmentDimensions

Description

Documentation

data SegmentDimensions

Specifies the dimension settings for a segment.

See: newSegmentDimensions smart constructor.

Constructors

SegmentDimensions'

Fields

metrics :: Maybe (HashMap Text MetricDimension)

One or more custom metrics to use as criteria for the segment.

location :: Maybe SegmentLocation

The location-based criteria, such as region or GPS coordinates, for the segment.

demographic :: Maybe SegmentDemographics

The demographic-based criteria, such as device platform, for the segment.

userAttributes :: Maybe (HashMap Text AttributeDimension)

One or more custom user attributes to use as criteria for the segment.

behavior :: Maybe SegmentBehaviors

The behavior-based criteria, such as how recently users have used your app, for the segment.

attributes :: Maybe (HashMap Text AttributeDimension)

One or more custom attributes to use as criteria for the segment.

newSegmentDimensions :: SegmentDimensions

Create a value of SegmentDimensions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metrics:SegmentDimensions', segmentDimensions_metrics - One or more custom metrics to use as criteria for the segment.

$sel:location:SegmentDimensions', segmentDimensions_location - The location-based criteria, such as region or GPS coordinates, for the segment.

$sel:demographic:SegmentDimensions', segmentDimensions_demographic - The demographic-based criteria, such as device platform, for the segment.

$sel:userAttributes:SegmentDimensions', segmentDimensions_userAttributes - One or more custom user attributes to use as criteria for the segment.

$sel:behavior:SegmentDimensions', segmentDimensions_behavior - The behavior-based criteria, such as how recently users have used your app, for the segment.

$sel:attributes:SegmentDimensions', segmentDimensions_attributes - One or more custom attributes to use as criteria for the segment.

segmentDimensions_metrics :: Lens' SegmentDimensions (Maybe (HashMap Text MetricDimension))

One or more custom metrics to use as criteria for the segment.

segmentDimensions_location :: Lens' SegmentDimensions (Maybe SegmentLocation)

The location-based criteria, such as region or GPS coordinates, for the segment.

segmentDimensions_demographic :: Lens' SegmentDimensions (Maybe SegmentDemographics)

The demographic-based criteria, such as device platform, for the segment.

segmentDimensions_userAttributes :: Lens' SegmentDimensions (Maybe (HashMap Text AttributeDimension))

One or more custom user attributes to use as criteria for the segment.

segmentDimensions_behavior :: Lens' SegmentDimensions (Maybe SegmentBehaviors)

The behavior-based criteria, such as how recently users have used your app, for the segment.

segmentDimensions_attributes :: Lens' SegmentDimensions (Maybe (HashMap Text AttributeDimension))

One or more custom attributes to use as criteria for the segment.