Amazonka.Pinpoint.Types.Session

Description

Documentation

data Session

Provides information about a session.

See: newSession smart constructor.

Constructors

Session'

Fields

stopTimestamp :: Maybe Text

The date and time when the session ended.

duration :: Maybe Int

The duration of the session, in milliseconds.

startTimestamp :: Text

The date and time when the session began.

id :: Text

The unique identifier for the session.

newSession

Arguments

:: Text
$sel:startTimestamp:Session'

-> Text
$sel:id:Session'

-> Session

Create a value of Session with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stopTimestamp:Session', session_stopTimestamp - The date and time when the session ended.

$sel:duration:Session', session_duration - The duration of the session, in milliseconds.

$sel:startTimestamp:Session', session_startTimestamp - The date and time when the session began.

$sel:id:Session', session_id - The unique identifier for the session.

session_stopTimestamp :: Lens' Session (Maybe Text)

The date and time when the session ended.

session_duration :: Lens' Session (Maybe Int)

The duration of the session, in milliseconds.

session_startTimestamp :: Lens' Session Text

The date and time when the session began.

session_id :: Lens' Session Text

The unique identifier for the session.