Amazonka.Pinpoint.Types.TemplateActiveVersionRequest

Description

Documentation

data TemplateActiveVersionRequest

Specifies which version of a message template to use as the active version of the template.

See: newTemplateActiveVersionRequest smart constructor.

Constructors

TemplateActiveVersionRequest'

Fields

version :: Maybe Text

The version of the message template to use as the active version of the template. Valid values are: latest, for the most recent version of the template; or, the unique identifier for any existing version of the template. If you specify an identifier, the value must match the identifier for an existing template version. To retrieve a list of versions and version identifiers for a template, use the Template Versions resource.

newTemplateActiveVersionRequest :: TemplateActiveVersionRequest

Create a value of TemplateActiveVersionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:TemplateActiveVersionRequest', templateActiveVersionRequest_version - The version of the message template to use as the active version of the template. Valid values are: latest, for the most recent version of the template; or, the unique identifier for any existing version of the template. If you specify an identifier, the value must match the identifier for an existing template version. To retrieve a list of versions and version identifiers for a template, use the Template Versions resource.

templateActiveVersionRequest_version :: Lens' TemplateActiveVersionRequest (Maybe Text)

The version of the message template to use as the active version of the template. Valid values are: latest, for the most recent version of the template; or, the unique identifier for any existing version of the template. If you specify an identifier, the value must match the identifier for an existing template version. To retrieve a list of versions and version identifiers for a template, use the Template Versions resource.