Amazonka.Pinpoint.UpdateApnsVoipChannel

Description

Enables the APNs VoIP channel for an application or updates the status and settings of the APNs VoIP channel for an application.

Creating a Request

data UpdateApnsVoipChannel

See: newUpdateApnsVoipChannel smart constructor.

Constructors

UpdateApnsVoipChannel'

Fields

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

aPNSVoipChannelRequest :: APNSVoipChannelRequest

newUpdateApnsVoipChannel

Arguments

:: Text
$sel:applicationId:UpdateApnsVoipChannel'

-> APNSVoipChannelRequest
$sel:aPNSVoipChannelRequest:UpdateApnsVoipChannel'

-> UpdateApnsVoipChannel

Create a value of UpdateApnsVoipChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:UpdateApnsVoipChannel', updateApnsVoipChannel_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

$sel:aPNSVoipChannelRequest:UpdateApnsVoipChannel', updateApnsVoipChannel_aPNSVoipChannelRequest - Undocumented member.

Request Lenses

updateApnsVoipChannel_applicationId :: Lens' UpdateApnsVoipChannel Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

updateApnsVoipChannel_aPNSVoipChannelRequest :: Lens' UpdateApnsVoipChannel APNSVoipChannelRequest

Undocumented member.

Destructuring the Response

data UpdateApnsVoipChannelResponse

See: newUpdateApnsVoipChannelResponse smart constructor.

Constructors

UpdateApnsVoipChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

aPNSVoipChannelResponse :: APNSVoipChannelResponse

newUpdateApnsVoipChannelResponse

Arguments

:: Int
$sel:httpStatus:UpdateApnsVoipChannelResponse'

-> APNSVoipChannelResponse
$sel:aPNSVoipChannelResponse:UpdateApnsVoipChannelResponse'

-> UpdateApnsVoipChannelResponse

Create a value of UpdateApnsVoipChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateApnsVoipChannelResponse', updateApnsVoipChannelResponse_httpStatus - The response's http status code.

$sel:aPNSVoipChannelResponse:UpdateApnsVoipChannelResponse', updateApnsVoipChannelResponse_aPNSVoipChannelResponse - Undocumented member.

Response Lenses

updateApnsVoipChannelResponse_httpStatus :: Lens' UpdateApnsVoipChannelResponse Int

The response's http status code.

updateApnsVoipChannelResponse_aPNSVoipChannelResponse :: Lens' UpdateApnsVoipChannelResponse APNSVoipChannelResponse

Undocumented member.