Amazonka.Pinpoint.UpdateEmailTemplate

Description

Updates an existing message template for messages that are sent through the email channel.

Creating a Request

data UpdateEmailTemplate

See: newUpdateEmailTemplate smart constructor.

Constructors

UpdateEmailTemplate'

Fields

version :: Maybe Text

The unique identifier for the version of the message template to update, retrieve information about, or delete. To retrieve identifiers and other information for all the versions of a template, use the Template Versions resource.

If specified, this value must match the identifier for an existing template version. If specified for an update operation, this value must match the identifier for the latest existing version of the template. This restriction helps ensure that race conditions don't occur.

If you don't specify a value for this parameter, Amazon Pinpoint does the following:

For a get operation, retrieves information about the active version of the template.

For an update operation, saves the updates to (overwrites) the latest existing version of the template, if the create-new-version parameter isn't used or is set to false.

For a delete operation, deletes the template, including all versions of the template.

createNewVersion :: Maybe Bool

Specifies whether to save the updates as a new version of the message template. Valid values are: true, save the updates as a new version; and, false, save the updates to (overwrite) the latest existing version of the template.

If you don't specify a value for this parameter, Amazon Pinpoint saves the updates to (overwrites) the latest existing version of the template. If you specify a value of true for this parameter, don't specify a value for the version parameter. Otherwise, an error will occur.

templateName :: Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

emailTemplateRequest :: EmailTemplateRequest

newUpdateEmailTemplate

Arguments

:: Text
$sel:templateName:UpdateEmailTemplate'

-> EmailTemplateRequest
$sel:emailTemplateRequest:UpdateEmailTemplate'

-> UpdateEmailTemplate

Create a value of UpdateEmailTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:UpdateEmailTemplate', updateEmailTemplate_version - The unique identifier for the version of the message template to update, retrieve information about, or delete. To retrieve identifiers and other information for all the versions of a template, use the Template Versions resource.

If specified, this value must match the identifier for an existing template version. If specified for an update operation, this value must match the identifier for the latest existing version of the template. This restriction helps ensure that race conditions don't occur.

If you don't specify a value for this parameter, Amazon Pinpoint does the following:

For a get operation, retrieves information about the active version of the template.

For an update operation, saves the updates to (overwrites) the latest existing version of the template, if the create-new-version parameter isn't used or is set to false.

For a delete operation, deletes the template, including all versions of the template.

$sel:createNewVersion:UpdateEmailTemplate', updateEmailTemplate_createNewVersion - Specifies whether to save the updates as a new version of the message template. Valid values are: true, save the updates as a new version; and, false, save the updates to (overwrite) the latest existing version of the template.

If you don't specify a value for this parameter, Amazon Pinpoint saves the updates to (overwrites) the latest existing version of the template. If you specify a value of true for this parameter, don't specify a value for the version parameter. Otherwise, an error will occur.

$sel:templateName:UpdateEmailTemplate', updateEmailTemplate_templateName - The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

$sel:emailTemplateRequest:UpdateEmailTemplate', updateEmailTemplate_emailTemplateRequest - Undocumented member.

Request Lenses

updateEmailTemplate_version :: Lens' UpdateEmailTemplate (Maybe Text)

The unique identifier for the version of the message template to update, retrieve information about, or delete. To retrieve identifiers and other information for all the versions of a template, use the Template Versions resource.

If specified, this value must match the identifier for an existing template version. If specified for an update operation, this value must match the identifier for the latest existing version of the template. This restriction helps ensure that race conditions don't occur.

If you don't specify a value for this parameter, Amazon Pinpoint does the following:

For a get operation, retrieves information about the active version of the template.

For an update operation, saves the updates to (overwrites) the latest existing version of the template, if the create-new-version parameter isn't used or is set to false.

For a delete operation, deletes the template, including all versions of the template.

updateEmailTemplate_createNewVersion :: Lens' UpdateEmailTemplate (Maybe Bool)

Specifies whether to save the updates as a new version of the message template. Valid values are: true, save the updates as a new version; and, false, save the updates to (overwrite) the latest existing version of the template.

If you don't specify a value for this parameter, Amazon Pinpoint saves the updates to (overwrites) the latest existing version of the template. If you specify a value of true for this parameter, don't specify a value for the version parameter. Otherwise, an error will occur.

updateEmailTemplate_templateName :: Lens' UpdateEmailTemplate Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

updateEmailTemplate_emailTemplateRequest :: Lens' UpdateEmailTemplate EmailTemplateRequest

Undocumented member.

Destructuring the Response

data UpdateEmailTemplateResponse

See: newUpdateEmailTemplateResponse smart constructor.

Constructors

UpdateEmailTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

messageBody :: MessageBody

newUpdateEmailTemplateResponse

Arguments

:: Int
$sel:httpStatus:UpdateEmailTemplateResponse'

-> MessageBody
$sel:messageBody:UpdateEmailTemplateResponse'

-> UpdateEmailTemplateResponse

Create a value of UpdateEmailTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateEmailTemplateResponse', updateEmailTemplateResponse_httpStatus - The response's http status code.

$sel:messageBody:UpdateEmailTemplateResponse', updateEmailTemplateResponse_messageBody - Undocumented member.

Response Lenses

updateEmailTemplateResponse_httpStatus :: Lens' UpdateEmailTemplateResponse Int

The response's http status code.

updateEmailTemplateResponse_messageBody :: Lens' UpdateEmailTemplateResponse MessageBody

Undocumented member.