Amazonka.Pinpoint.UpdateGcmChannel

Description

Enables the GCM channel for an application or updates the status and settings of the GCM channel for an application.

Creating a Request

data UpdateGcmChannel

See: newUpdateGcmChannel smart constructor.

Constructors

UpdateGcmChannel'

Fields

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

gCMChannelRequest :: GCMChannelRequest

newUpdateGcmChannel

Arguments

:: Text
$sel:applicationId:UpdateGcmChannel'

-> GCMChannelRequest
$sel:gCMChannelRequest:UpdateGcmChannel'

-> UpdateGcmChannel

Create a value of UpdateGcmChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:UpdateGcmChannel', updateGcmChannel_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

$sel:gCMChannelRequest:UpdateGcmChannel', updateGcmChannel_gCMChannelRequest - Undocumented member.

Request Lenses

updateGcmChannel_applicationId :: Lens' UpdateGcmChannel Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

updateGcmChannel_gCMChannelRequest :: Lens' UpdateGcmChannel GCMChannelRequest

Undocumented member.

Destructuring the Response

data UpdateGcmChannelResponse

See: newUpdateGcmChannelResponse smart constructor.

Constructors

UpdateGcmChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

gCMChannelResponse :: GCMChannelResponse

newUpdateGcmChannelResponse

Arguments

:: Int
$sel:httpStatus:UpdateGcmChannelResponse'

-> GCMChannelResponse
$sel:gCMChannelResponse:UpdateGcmChannelResponse'

-> UpdateGcmChannelResponse

Create a value of UpdateGcmChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateGcmChannelResponse', updateGcmChannelResponse_httpStatus - The response's http status code.

$sel:gCMChannelResponse:UpdateGcmChannelResponse', updateGcmChannelResponse_gCMChannelResponse - Undocumented member.

Response Lenses

updateGcmChannelResponse_httpStatus :: Lens' UpdateGcmChannelResponse Int

The response's http status code.

updateGcmChannelResponse_gCMChannelResponse :: Lens' UpdateGcmChannelResponse GCMChannelResponse

Undocumented member.