Amazonka.Pinpoint.UpdateInAppTemplate

Description

Updates an existing message template for messages sent through the in-app message channel.

Creating a Request

data UpdateInAppTemplate

See: newUpdateInAppTemplate smart constructor.

Constructors

UpdateInAppTemplate'

Fields

version :: Maybe Text

The unique identifier for the version of the message template to update, retrieve information about, or delete. To retrieve identifiers and other information for all the versions of a template, use the Template Versions resource.

If specified, this value must match the identifier for an existing template version. If specified for an update operation, this value must match the identifier for the latest existing version of the template. This restriction helps ensure that race conditions don't occur.

If you don't specify a value for this parameter, Amazon Pinpoint does the following:

For a get operation, retrieves information about the active version of the template.

For an update operation, saves the updates to (overwrites) the latest existing version of the template, if the create-new-version parameter isn't used or is set to false.

For a delete operation, deletes the template, including all versions of the template.

createNewVersion :: Maybe Bool

Specifies whether to save the updates as a new version of the message template. Valid values are: true, save the updates as a new version; and, false, save the updates to (overwrite) the latest existing version of the template.

If you don't specify a value for this parameter, Amazon Pinpoint saves the updates to (overwrites) the latest existing version of the template. If you specify a value of true for this parameter, don't specify a value for the version parameter. Otherwise, an error will occur.

templateName :: Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

inAppTemplateRequest :: InAppTemplateRequest

newUpdateInAppTemplate

Arguments

:: Text
$sel:templateName:UpdateInAppTemplate'

-> InAppTemplateRequest
$sel:inAppTemplateRequest:UpdateInAppTemplate'

-> UpdateInAppTemplate

Create a value of UpdateInAppTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:UpdateInAppTemplate', updateInAppTemplate_version - The unique identifier for the version of the message template to update, retrieve information about, or delete. To retrieve identifiers and other information for all the versions of a template, use the Template Versions resource.

If specified, this value must match the identifier for an existing template version. If specified for an update operation, this value must match the identifier for the latest existing version of the template. This restriction helps ensure that race conditions don't occur.

If you don't specify a value for this parameter, Amazon Pinpoint does the following:

For a get operation, retrieves information about the active version of the template.

For an update operation, saves the updates to (overwrites) the latest existing version of the template, if the create-new-version parameter isn't used or is set to false.

For a delete operation, deletes the template, including all versions of the template.

$sel:createNewVersion:UpdateInAppTemplate', updateInAppTemplate_createNewVersion - Specifies whether to save the updates as a new version of the message template. Valid values are: true, save the updates as a new version; and, false, save the updates to (overwrite) the latest existing version of the template.

If you don't specify a value for this parameter, Amazon Pinpoint saves the updates to (overwrites) the latest existing version of the template. If you specify a value of true for this parameter, don't specify a value for the version parameter. Otherwise, an error will occur.

$sel:templateName:UpdateInAppTemplate', updateInAppTemplate_templateName - The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

$sel:inAppTemplateRequest:UpdateInAppTemplate', updateInAppTemplate_inAppTemplateRequest - Undocumented member.

Request Lenses

updateInAppTemplate_version :: Lens' UpdateInAppTemplate (Maybe Text)

The unique identifier for the version of the message template to update, retrieve information about, or delete. To retrieve identifiers and other information for all the versions of a template, use the Template Versions resource.

If specified, this value must match the identifier for an existing template version. If specified for an update operation, this value must match the identifier for the latest existing version of the template. This restriction helps ensure that race conditions don't occur.

If you don't specify a value for this parameter, Amazon Pinpoint does the following:

For a get operation, retrieves information about the active version of the template.

For an update operation, saves the updates to (overwrites) the latest existing version of the template, if the create-new-version parameter isn't used or is set to false.

For a delete operation, deletes the template, including all versions of the template.

updateInAppTemplate_createNewVersion :: Lens' UpdateInAppTemplate (Maybe Bool)

Specifies whether to save the updates as a new version of the message template. Valid values are: true, save the updates as a new version; and, false, save the updates to (overwrite) the latest existing version of the template.

If you don't specify a value for this parameter, Amazon Pinpoint saves the updates to (overwrites) the latest existing version of the template. If you specify a value of true for this parameter, don't specify a value for the version parameter. Otherwise, an error will occur.

updateInAppTemplate_templateName :: Lens' UpdateInAppTemplate Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

updateInAppTemplate_inAppTemplateRequest :: Lens' UpdateInAppTemplate InAppTemplateRequest

Undocumented member.

Destructuring the Response

data UpdateInAppTemplateResponse

See: newUpdateInAppTemplateResponse smart constructor.

Constructors

UpdateInAppTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

messageBody :: MessageBody

newUpdateInAppTemplateResponse

Arguments

:: Int
$sel:httpStatus:UpdateInAppTemplateResponse'

-> MessageBody
$sel:messageBody:UpdateInAppTemplateResponse'

-> UpdateInAppTemplateResponse

Create a value of UpdateInAppTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateInAppTemplateResponse', updateInAppTemplateResponse_httpStatus - The response's http status code.

$sel:messageBody:UpdateInAppTemplateResponse', updateInAppTemplateResponse_messageBody - Undocumented member.

Response Lenses

updateInAppTemplateResponse_httpStatus :: Lens' UpdateInAppTemplateResponse Int

The response's http status code.

updateInAppTemplateResponse_messageBody :: Lens' UpdateInAppTemplateResponse MessageBody

Undocumented member.