Amazonka.Pinpoint.UpdatePushTemplate

Description

Updates an existing message template for messages that are sent through a push notification channel.

Creating a Request

data UpdatePushTemplate

See: newUpdatePushTemplate smart constructor.

Constructors

UpdatePushTemplate'

Fields

version :: Maybe Text

The unique identifier for the version of the message template to update, retrieve information about, or delete. To retrieve identifiers and other information for all the versions of a template, use the Template Versions resource.

If specified, this value must match the identifier for an existing template version. If specified for an update operation, this value must match the identifier for the latest existing version of the template. This restriction helps ensure that race conditions don't occur.

If you don't specify a value for this parameter, Amazon Pinpoint does the following:

For a get operation, retrieves information about the active version of the template.

For an update operation, saves the updates to (overwrites) the latest existing version of the template, if the create-new-version parameter isn't used or is set to false.

For a delete operation, deletes the template, including all versions of the template.

createNewVersion :: Maybe Bool

Specifies whether to save the updates as a new version of the message template. Valid values are: true, save the updates as a new version; and, false, save the updates to (overwrite) the latest existing version of the template.

If you don't specify a value for this parameter, Amazon Pinpoint saves the updates to (overwrites) the latest existing version of the template. If you specify a value of true for this parameter, don't specify a value for the version parameter. Otherwise, an error will occur.

templateName :: Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

pushNotificationTemplateRequest :: PushNotificationTemplateRequest

newUpdatePushTemplate

Arguments

:: Text
$sel:templateName:UpdatePushTemplate'

-> PushNotificationTemplateRequest
$sel:pushNotificationTemplateRequest:UpdatePushTemplate'

-> UpdatePushTemplate

Create a value of UpdatePushTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:UpdatePushTemplate', updatePushTemplate_version - The unique identifier for the version of the message template to update, retrieve information about, or delete. To retrieve identifiers and other information for all the versions of a template, use the Template Versions resource.

If specified, this value must match the identifier for an existing template version. If specified for an update operation, this value must match the identifier for the latest existing version of the template. This restriction helps ensure that race conditions don't occur.

If you don't specify a value for this parameter, Amazon Pinpoint does the following:

For a get operation, retrieves information about the active version of the template.

For an update operation, saves the updates to (overwrites) the latest existing version of the template, if the create-new-version parameter isn't used or is set to false.

For a delete operation, deletes the template, including all versions of the template.

$sel:createNewVersion:UpdatePushTemplate', updatePushTemplate_createNewVersion - Specifies whether to save the updates as a new version of the message template. Valid values are: true, save the updates as a new version; and, false, save the updates to (overwrite) the latest existing version of the template.

If you don't specify a value for this parameter, Amazon Pinpoint saves the updates to (overwrites) the latest existing version of the template. If you specify a value of true for this parameter, don't specify a value for the version parameter. Otherwise, an error will occur.

$sel:templateName:UpdatePushTemplate', updatePushTemplate_templateName - The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

$sel:pushNotificationTemplateRequest:UpdatePushTemplate', updatePushTemplate_pushNotificationTemplateRequest - Undocumented member.

Request Lenses

updatePushTemplate_version :: Lens' UpdatePushTemplate (Maybe Text)

The unique identifier for the version of the message template to update, retrieve information about, or delete. To retrieve identifiers and other information for all the versions of a template, use the Template Versions resource.

If specified, this value must match the identifier for an existing template version. If specified for an update operation, this value must match the identifier for the latest existing version of the template. This restriction helps ensure that race conditions don't occur.

If you don't specify a value for this parameter, Amazon Pinpoint does the following:

For a get operation, retrieves information about the active version of the template.

For an update operation, saves the updates to (overwrites) the latest existing version of the template, if the create-new-version parameter isn't used or is set to false.

For a delete operation, deletes the template, including all versions of the template.

updatePushTemplate_createNewVersion :: Lens' UpdatePushTemplate (Maybe Bool)

Specifies whether to save the updates as a new version of the message template. Valid values are: true, save the updates as a new version; and, false, save the updates to (overwrite) the latest existing version of the template.

If you don't specify a value for this parameter, Amazon Pinpoint saves the updates to (overwrites) the latest existing version of the template. If you specify a value of true for this parameter, don't specify a value for the version parameter. Otherwise, an error will occur.

updatePushTemplate_templateName :: Lens' UpdatePushTemplate Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

updatePushTemplate_pushNotificationTemplateRequest :: Lens' UpdatePushTemplate PushNotificationTemplateRequest

Undocumented member.

Destructuring the Response

data UpdatePushTemplateResponse

See: newUpdatePushTemplateResponse smart constructor.

Constructors

UpdatePushTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

messageBody :: MessageBody

newUpdatePushTemplateResponse

Arguments

:: Int
$sel:httpStatus:UpdatePushTemplateResponse'

-> MessageBody
$sel:messageBody:UpdatePushTemplateResponse'

-> UpdatePushTemplateResponse

Create a value of UpdatePushTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdatePushTemplateResponse', updatePushTemplateResponse_httpStatus - The response's http status code.

$sel:messageBody:UpdatePushTemplateResponse', updatePushTemplateResponse_messageBody - Undocumented member.

Response Lenses

updatePushTemplateResponse_httpStatus :: Lens' UpdatePushTemplateResponse Int

The response's http status code.

updatePushTemplateResponse_messageBody :: Lens' UpdatePushTemplateResponse MessageBody

Undocumented member.