Amazonka.Pinpoint.UpdateTemplateActiveVersion

Description

Changes the status of a specific version of a message template to active.

Creating a Request

data UpdateTemplateActiveVersion

See: newUpdateTemplateActiveVersion smart constructor.

Constructors

UpdateTemplateActiveVersion'

Fields

templateName :: Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

templateType :: Text

The type of channel that the message template is designed for. Valid values are: EMAIL, PUSH, SMS, and VOICE.

templateActiveVersionRequest :: TemplateActiveVersionRequest

newUpdateTemplateActiveVersion

Arguments

:: Text
$sel:templateName:UpdateTemplateActiveVersion'

-> Text
$sel:templateType:UpdateTemplateActiveVersion'

-> TemplateActiveVersionRequest
$sel:templateActiveVersionRequest:UpdateTemplateActiveVersion'

-> UpdateTemplateActiveVersion

Create a value of UpdateTemplateActiveVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:UpdateTemplateActiveVersion', updateTemplateActiveVersion_templateName - The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

$sel:templateType:UpdateTemplateActiveVersion', updateTemplateActiveVersion_templateType - The type of channel that the message template is designed for. Valid values are: EMAIL, PUSH, SMS, and VOICE.

$sel:templateActiveVersionRequest:UpdateTemplateActiveVersion', updateTemplateActiveVersion_templateActiveVersionRequest - Undocumented member.

Request Lenses

updateTemplateActiveVersion_templateName :: Lens' UpdateTemplateActiveVersion Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

updateTemplateActiveVersion_templateType :: Lens' UpdateTemplateActiveVersion Text

The type of channel that the message template is designed for. Valid values are: EMAIL, PUSH, SMS, and VOICE.

updateTemplateActiveVersion_templateActiveVersionRequest :: Lens' UpdateTemplateActiveVersion TemplateActiveVersionRequest

Undocumented member.

Destructuring the Response

data UpdateTemplateActiveVersionResponse

See: newUpdateTemplateActiveVersionResponse smart constructor.

Constructors

UpdateTemplateActiveVersionResponse'

Fields

httpStatus :: Int

The response's http status code.

messageBody :: MessageBody

newUpdateTemplateActiveVersionResponse

Arguments

:: Int
$sel:httpStatus:UpdateTemplateActiveVersionResponse'

-> MessageBody
$sel:messageBody:UpdateTemplateActiveVersionResponse'

-> UpdateTemplateActiveVersionResponse

Create a value of UpdateTemplateActiveVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateTemplateActiveVersionResponse', updateTemplateActiveVersionResponse_httpStatus - The response's http status code.

$sel:messageBody:UpdateTemplateActiveVersionResponse', updateTemplateActiveVersionResponse_messageBody - Undocumented member.

Response Lenses

updateTemplateActiveVersionResponse_httpStatus :: Lens' UpdateTemplateActiveVersionResponse Int

The response's http status code.

updateTemplateActiveVersionResponse_messageBody :: Lens' UpdateTemplateActiveVersionResponse MessageBody

Undocumented member.