Amazonka.PinpointEmail.CreateEmailIdentity

Description

Verifies an email identity for use with Amazon Pinpoint. In Amazon Pinpoint, an identity is an email address or domain that you use when you send email. Before you can use an identity to send email with Amazon Pinpoint, you first have to verify it. By verifying an address, you demonstrate that you're the owner of the address, and that you've given Amazon Pinpoint permission to send email from the address.

When you verify an email address, Amazon Pinpoint sends an email to the address. Your email address is verified as soon as you follow the link in the verification email.

When you verify a domain, this operation provides a set of DKIM tokens, which you can convert into CNAME tokens. You add these CNAME tokens to the DNS configuration for your domain. Your domain is verified when Amazon Pinpoint detects these records in the DNS configuration for your domain. It usually takes around 72 hours to complete the domain verification process.

Creating a Request

data CreateEmailIdentity

A request to begin the verification process for an email identity (an email address or domain).

See: newCreateEmailIdentity smart constructor.

Constructors

CreateEmailIdentity'

Fields

tags :: Maybe [Tag]

An array of objects that define the tags (keys and values) that you want to associate with the email identity.

emailIdentity :: Text

The email address or domain that you want to verify.

newCreateEmailIdentity

Arguments

:: Text
$sel:emailIdentity:CreateEmailIdentity'

-> CreateEmailIdentity

Create a value of CreateEmailIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateEmailIdentity', createEmailIdentity_tags - An array of objects that define the tags (keys and values) that you want to associate with the email identity.

$sel:emailIdentity:CreateEmailIdentity', createEmailIdentity_emailIdentity - The email address or domain that you want to verify.

Request Lenses

createEmailIdentity_tags :: Lens' CreateEmailIdentity (Maybe [Tag])

An array of objects that define the tags (keys and values) that you want to associate with the email identity.

createEmailIdentity_emailIdentity :: Lens' CreateEmailIdentity Text

The email address or domain that you want to verify.

Destructuring the Response

data CreateEmailIdentityResponse

If the email identity is a domain, this object contains tokens that you can use to create a set of CNAME records. To sucessfully verify your domain, you have to add these records to the DNS configuration for your domain.

If the email identity is an email address, this object is empty.

See: newCreateEmailIdentityResponse smart constructor.

Constructors

CreateEmailIdentityResponse'

Fields

dkimAttributes :: Maybe DkimAttributes

An object that contains information about the DKIM attributes for the identity. This object includes the tokens that you use to create the CNAME records that are required to complete the DKIM verification process.

verifiedForSendingStatus :: Maybe Bool

Specifies whether or not the identity is verified. In Amazon Pinpoint, you can only send email from verified email addresses or domains. For more information about verifying identities, see the Amazon Pinpoint User Guide.

identityType :: Maybe IdentityType

The email identity type.

httpStatus :: Int

The response's http status code.

newCreateEmailIdentityResponse

Arguments

:: Int
$sel:httpStatus:CreateEmailIdentityResponse'

-> CreateEmailIdentityResponse

Create a value of CreateEmailIdentityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dkimAttributes:CreateEmailIdentityResponse', createEmailIdentityResponse_dkimAttributes - An object that contains information about the DKIM attributes for the identity. This object includes the tokens that you use to create the CNAME records that are required to complete the DKIM verification process.

$sel:verifiedForSendingStatus:CreateEmailIdentityResponse', createEmailIdentityResponse_verifiedForSendingStatus - Specifies whether or not the identity is verified. In Amazon Pinpoint, you can only send email from verified email addresses or domains. For more information about verifying identities, see the Amazon Pinpoint User Guide.

$sel:identityType:CreateEmailIdentityResponse', createEmailIdentityResponse_identityType - The email identity type.

$sel:httpStatus:CreateEmailIdentityResponse', createEmailIdentityResponse_httpStatus - The response's http status code.

Response Lenses

createEmailIdentityResponse_dkimAttributes :: Lens' CreateEmailIdentityResponse (Maybe DkimAttributes)

An object that contains information about the DKIM attributes for the identity. This object includes the tokens that you use to create the CNAME records that are required to complete the DKIM verification process.

createEmailIdentityResponse_verifiedForSendingStatus :: Lens' CreateEmailIdentityResponse (Maybe Bool)

Specifies whether or not the identity is verified. In Amazon Pinpoint, you can only send email from verified email addresses or domains. For more information about verifying identities, see the Amazon Pinpoint User Guide.

createEmailIdentityResponse_identityType :: Lens' CreateEmailIdentityResponse (Maybe IdentityType)

The email identity type.

createEmailIdentityResponse_httpStatus :: Lens' CreateEmailIdentityResponse Int

The response's http status code.