Amazonka.PinpointEmail.PutDedicatedIpWarmupAttributes

Description

Creating a Request

data PutDedicatedIpWarmupAttributes

A request to change the warm-up attributes for a dedicated IP address. This operation is useful when you want to resume the warm-up process for an existing IP address.

See: newPutDedicatedIpWarmupAttributes smart constructor.

Constructors

PutDedicatedIpWarmupAttributes'

Fields

ip :: Text

The dedicated IP address that you want to update the warm-up attributes for.

warmupPercentage :: Int

The warm-up percentage that you want to associate with the dedicated IP address.

newPutDedicatedIpWarmupAttributes

Arguments

:: Text
$sel:ip:PutDedicatedIpWarmupAttributes'

-> Int
$sel:warmupPercentage:PutDedicatedIpWarmupAttributes'

-> PutDedicatedIpWarmupAttributes

Create a value of PutDedicatedIpWarmupAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ip:PutDedicatedIpWarmupAttributes', putDedicatedIpWarmupAttributes_ip - The dedicated IP address that you want to update the warm-up attributes for.

$sel:warmupPercentage:PutDedicatedIpWarmupAttributes', putDedicatedIpWarmupAttributes_warmupPercentage - The warm-up percentage that you want to associate with the dedicated IP address.

Request Lenses

putDedicatedIpWarmupAttributes_ip :: Lens' PutDedicatedIpWarmupAttributes Text

The dedicated IP address that you want to update the warm-up attributes for.

putDedicatedIpWarmupAttributes_warmupPercentage :: Lens' PutDedicatedIpWarmupAttributes Int

The warm-up percentage that you want to associate with the dedicated IP address.

Destructuring the Response

data PutDedicatedIpWarmupAttributesResponse

An HTTP 200 response if the request succeeds, or an error message if the request fails.

See: newPutDedicatedIpWarmupAttributesResponse smart constructor.

Constructors

PutDedicatedIpWarmupAttributesResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutDedicatedIpWarmupAttributesResponse

Arguments

:: Int
$sel:httpStatus:PutDedicatedIpWarmupAttributesResponse'

-> PutDedicatedIpWarmupAttributesResponse

Create a value of PutDedicatedIpWarmupAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutDedicatedIpWarmupAttributesResponse', putDedicatedIpWarmupAttributesResponse_httpStatus - The response's http status code.

Response Lenses

putDedicatedIpWarmupAttributesResponse_httpStatus :: Lens' PutDedicatedIpWarmupAttributesResponse Int

The response's http status code.