Amazonka.PinpointSMSVoice

Description

Derived from API version 2018-09-05 of the AWS service descriptions, licensed under Apache 2.0.

Pinpoint SMS and Voice Messaging public facing APIs

Service Configuration

defaultService :: Service

API version 2018-09-05 of the Amazon Pinpoint SMS and Voice Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by PinpointSMSVoice.

NotFoundException

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource you attempted to access doesn't exist.

TooManyRequestsException

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

You've issued too many requests to the resource. Wait a few minutes, and then try again.

InternalServiceErrorException

_InternalServiceErrorException :: AsError a => Getting (First ServiceError) a ServiceError

The API encountered an unexpected error and couldn't complete the request. You might be able to successfully issue the request again in the future.

BadRequestException

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The input you provided is invalid.

AlreadyExistsException

_AlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource specified in your request already exists.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

There are too many instances of the specified resource type.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

UpdateConfigurationSetEventDestination

data UpdateConfigurationSetEventDestination

UpdateConfigurationSetEventDestinationRequest

See: newUpdateConfigurationSetEventDestination smart constructor.

Constructors

UpdateConfigurationSetEventDestination' (Maybe EventDestinationDefinition) Text Text

newUpdateConfigurationSetEventDestination

Arguments

:: Text
$sel:eventDestinationName:UpdateConfigurationSetEventDestination'

-> Text
$sel:configurationSetName:UpdateConfigurationSetEventDestination'

-> UpdateConfigurationSetEventDestination

Create a value of UpdateConfigurationSetEventDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventDestination:UpdateConfigurationSetEventDestination', updateConfigurationSetEventDestination_eventDestination - Undocumented member.

$sel:eventDestinationName:UpdateConfigurationSetEventDestination', updateConfigurationSetEventDestination_eventDestinationName - EventDestinationName

$sel:configurationSetName:UpdateConfigurationSetEventDestination', updateConfigurationSetEventDestination_configurationSetName - ConfigurationSetName

data UpdateConfigurationSetEventDestinationResponse

An empty object that indicates that the event destination was updated successfully.

See: newUpdateConfigurationSetEventDestinationResponse smart constructor.

Constructors

UpdateConfigurationSetEventDestinationResponse' Int

newUpdateConfigurationSetEventDestinationResponse

Arguments

:: Int
$sel:httpStatus:UpdateConfigurationSetEventDestinationResponse'

-> UpdateConfigurationSetEventDestinationResponse

Create a value of UpdateConfigurationSetEventDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateConfigurationSetEventDestinationResponse', updateConfigurationSetEventDestinationResponse_httpStatus - The response's http status code.

DeleteConfigurationSetEventDestination

data DeleteConfigurationSetEventDestination

See: newDeleteConfigurationSetEventDestination smart constructor.

Constructors

DeleteConfigurationSetEventDestination' Text Text

newDeleteConfigurationSetEventDestination

Arguments

:: Text
$sel:eventDestinationName:DeleteConfigurationSetEventDestination'

-> Text
$sel:configurationSetName:DeleteConfigurationSetEventDestination'

-> DeleteConfigurationSetEventDestination

Create a value of DeleteConfigurationSetEventDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventDestinationName:DeleteConfigurationSetEventDestination', deleteConfigurationSetEventDestination_eventDestinationName - EventDestinationName

$sel:configurationSetName:DeleteConfigurationSetEventDestination', deleteConfigurationSetEventDestination_configurationSetName - ConfigurationSetName

data DeleteConfigurationSetEventDestinationResponse

An empty object that indicates that the event destination was deleted successfully.

See: newDeleteConfigurationSetEventDestinationResponse smart constructor.

Constructors

DeleteConfigurationSetEventDestinationResponse' Int

newDeleteConfigurationSetEventDestinationResponse

Arguments

:: Int
$sel:httpStatus:DeleteConfigurationSetEventDestinationResponse'

-> DeleteConfigurationSetEventDestinationResponse

Create a value of DeleteConfigurationSetEventDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteConfigurationSetEventDestinationResponse', deleteConfigurationSetEventDestinationResponse_httpStatus - The response's http status code.

DeleteConfigurationSet

data DeleteConfigurationSet

See: newDeleteConfigurationSet smart constructor.

Constructors

DeleteConfigurationSet' Text

newDeleteConfigurationSet

Arguments

:: Text
$sel:configurationSetName:DeleteConfigurationSet'

-> DeleteConfigurationSet

Create a value of DeleteConfigurationSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetName:DeleteConfigurationSet', deleteConfigurationSet_configurationSetName - ConfigurationSetName

data DeleteConfigurationSetResponse

An empty object that indicates that the configuration set was deleted successfully.

See: newDeleteConfigurationSetResponse smart constructor.

Constructors

DeleteConfigurationSetResponse' Int

newDeleteConfigurationSetResponse

Arguments

:: Int
$sel:httpStatus:DeleteConfigurationSetResponse'

-> DeleteConfigurationSetResponse

Create a value of DeleteConfigurationSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteConfigurationSetResponse', deleteConfigurationSetResponse_httpStatus - The response's http status code.

SendVoiceMessage

data SendVoiceMessage

SendVoiceMessageRequest

See: newSendVoiceMessage smart constructor.

Constructors

SendVoiceMessage' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe VoiceMessageContent) (Maybe Text)

newSendVoiceMessage :: SendVoiceMessage

Create a value of SendVoiceMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetName:SendVoiceMessage', sendVoiceMessage_configurationSetName - The name of the configuration set that you want to use to send the message.

$sel:callerId:SendVoiceMessage', sendVoiceMessage_callerId - The phone number that appears on recipients' devices when they receive the message.

$sel:originationPhoneNumber:SendVoiceMessage', sendVoiceMessage_originationPhoneNumber - The phone number that Amazon Pinpoint should use to send the voice message. This isn't necessarily the phone number that appears on recipients' devices when they receive the message, because you can specify a CallerId parameter in the request.

$sel:content:SendVoiceMessage', sendVoiceMessage_content - Undocumented member.

$sel:destinationPhoneNumber:SendVoiceMessage', sendVoiceMessage_destinationPhoneNumber - The phone number that you want to send the voice message to.

data SendVoiceMessageResponse

An object that that contains the Message ID of a Voice message that was sent successfully.

See: newSendVoiceMessageResponse smart constructor.

Constructors

SendVoiceMessageResponse' (Maybe Text) Int

newSendVoiceMessageResponse

Arguments

:: Int
$sel:httpStatus:SendVoiceMessageResponse'

-> SendVoiceMessageResponse

Create a value of SendVoiceMessageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageId:SendVoiceMessageResponse', sendVoiceMessageResponse_messageId - A unique identifier for the voice message.

$sel:httpStatus:SendVoiceMessageResponse', sendVoiceMessageResponse_httpStatus - The response's http status code.

GetConfigurationSetEventDestinations

data GetConfigurationSetEventDestinations

See: newGetConfigurationSetEventDestinations smart constructor.

Constructors

GetConfigurationSetEventDestinations' Text

newGetConfigurationSetEventDestinations

Arguments

:: Text
$sel:configurationSetName:GetConfigurationSetEventDestinations'

-> GetConfigurationSetEventDestinations

Create a value of GetConfigurationSetEventDestinations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetName:GetConfigurationSetEventDestinations', getConfigurationSetEventDestinations_configurationSetName - ConfigurationSetName

data GetConfigurationSetEventDestinationsResponse

An object that contains information about an event destination.

See: newGetConfigurationSetEventDestinationsResponse smart constructor.

Constructors

GetConfigurationSetEventDestinationsResponse' (Maybe [EventDestination]) Int

newGetConfigurationSetEventDestinationsResponse

Arguments

:: Int
$sel:httpStatus:GetConfigurationSetEventDestinationsResponse'

-> GetConfigurationSetEventDestinationsResponse

Create a value of GetConfigurationSetEventDestinationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventDestinations:GetConfigurationSetEventDestinationsResponse', getConfigurationSetEventDestinationsResponse_eventDestinations - Undocumented member.

$sel:httpStatus:GetConfigurationSetEventDestinationsResponse', getConfigurationSetEventDestinationsResponse_httpStatus - The response's http status code.

CreateConfigurationSetEventDestination

data CreateConfigurationSetEventDestination

Create a new event destination in a configuration set.

See: newCreateConfigurationSetEventDestination smart constructor.

Constructors

CreateConfigurationSetEventDestination' (Maybe EventDestinationDefinition) (Maybe Text) Text

newCreateConfigurationSetEventDestination

Arguments

:: Text
$sel:configurationSetName:CreateConfigurationSetEventDestination'

-> CreateConfigurationSetEventDestination

Create a value of CreateConfigurationSetEventDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventDestination:CreateConfigurationSetEventDestination', createConfigurationSetEventDestination_eventDestination - Undocumented member.

$sel:eventDestinationName:CreateConfigurationSetEventDestination', createConfigurationSetEventDestination_eventDestinationName - A name that identifies the event destination.

$sel:configurationSetName:CreateConfigurationSetEventDestination', createConfigurationSetEventDestination_configurationSetName - ConfigurationSetName

data CreateConfigurationSetEventDestinationResponse

An empty object that indicates that the event destination was created successfully.

See: newCreateConfigurationSetEventDestinationResponse smart constructor.

Constructors

CreateConfigurationSetEventDestinationResponse' Int

newCreateConfigurationSetEventDestinationResponse

Arguments

:: Int
$sel:httpStatus:CreateConfigurationSetEventDestinationResponse'

-> CreateConfigurationSetEventDestinationResponse

Create a value of CreateConfigurationSetEventDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateConfigurationSetEventDestinationResponse', createConfigurationSetEventDestinationResponse_httpStatus - The response's http status code.

CreateConfigurationSet

data CreateConfigurationSet

A request to create a new configuration set.

See: newCreateConfigurationSet smart constructor.

Constructors

CreateConfigurationSet' (Maybe Text)

newCreateConfigurationSet :: CreateConfigurationSet

Create a value of CreateConfigurationSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetName:CreateConfigurationSet', createConfigurationSet_configurationSetName - The name that you want to give the configuration set.

data CreateConfigurationSetResponse

An empty object that indicates that the configuration set was successfully created.

See: newCreateConfigurationSetResponse smart constructor.

Constructors

CreateConfigurationSetResponse' Int

newCreateConfigurationSetResponse

Arguments

:: Int
$sel:httpStatus:CreateConfigurationSetResponse'

-> CreateConfigurationSetResponse

Create a value of CreateConfigurationSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateConfigurationSetResponse', createConfigurationSetResponse_httpStatus - The response's http status code.

Types

EventType

newtype EventType

The types of events that are sent to the event destination.

Constructors

EventType'

Fields

fromEventType :: Text

Bundled Patterns

pattern EventType_ANSWERED :: EventType

pattern EventType_BUSY :: EventType

pattern EventType_COMPLETED_CALL :: EventType

pattern EventType_FAILED :: EventType

pattern EventType_INITIATED_CALL :: EventType

pattern EventType_NO_ANSWER :: EventType

pattern EventType_RINGING :: EventType

CallInstructionsMessageType

data CallInstructionsMessageType

An object that defines a message that contains text formatted using Amazon Pinpoint Voice Instructions markup.

See: newCallInstructionsMessageType smart constructor.

Constructors

CallInstructionsMessageType' (Maybe Text)

newCallInstructionsMessageType :: CallInstructionsMessageType

Create a value of CallInstructionsMessageType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:text:CallInstructionsMessageType', callInstructionsMessageType_text - The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

CloudWatchLogsDestination

data CloudWatchLogsDestination

An object that contains information about an event destination that sends data to Amazon CloudWatch Logs.

See: newCloudWatchLogsDestination smart constructor.

Constructors

CloudWatchLogsDestination' (Maybe Text) (Maybe Text)

newCloudWatchLogsDestination :: CloudWatchLogsDestination

Create a value of CloudWatchLogsDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamRoleArn:CloudWatchLogsDestination', cloudWatchLogsDestination_iamRoleArn - The Amazon Resource Name (ARN) of an Amazon Identity and Access Management (IAM) role that is able to write event data to an Amazon CloudWatch destination.

$sel:logGroupArn:CloudWatchLogsDestination', cloudWatchLogsDestination_logGroupArn - The name of the Amazon CloudWatch Log Group that you want to record events in.

EventDestination

data EventDestination

An object that defines an event destination.

See: newEventDestination smart constructor.

Constructors

EventDestination' (Maybe [EventType]) (Maybe Bool) (Maybe KinesisFirehoseDestination) (Maybe Text) (Maybe SnsDestination) (Maybe CloudWatchLogsDestination)

newEventDestination :: EventDestination

Create a value of EventDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:matchingEventTypes:EventDestination', eventDestination_matchingEventTypes - Undocumented member.

$sel:enabled:EventDestination', eventDestination_enabled - Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

$sel:kinesisFirehoseDestination:EventDestination', eventDestination_kinesisFirehoseDestination - Undocumented member.

$sel:name:EventDestination', eventDestination_name - A name that identifies the event destination configuration.

$sel:snsDestination:EventDestination', eventDestination_snsDestination - Undocumented member.

$sel:cloudWatchLogsDestination:EventDestination', eventDestination_cloudWatchLogsDestination - Undocumented member.

EventDestinationDefinition

data EventDestinationDefinition

An object that defines a single event destination.

See: newEventDestinationDefinition smart constructor.

Constructors

EventDestinationDefinition' (Maybe [EventType]) (Maybe Bool) (Maybe KinesisFirehoseDestination) (Maybe SnsDestination) (Maybe CloudWatchLogsDestination)

newEventDestinationDefinition :: EventDestinationDefinition

Create a value of EventDestinationDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:matchingEventTypes:EventDestinationDefinition', eventDestinationDefinition_matchingEventTypes - Undocumented member.

$sel:enabled:EventDestinationDefinition', eventDestinationDefinition_enabled - Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

$sel:kinesisFirehoseDestination:EventDestinationDefinition', eventDestinationDefinition_kinesisFirehoseDestination - Undocumented member.

$sel:snsDestination:EventDestinationDefinition', eventDestinationDefinition_snsDestination - Undocumented member.

$sel:cloudWatchLogsDestination:EventDestinationDefinition', eventDestinationDefinition_cloudWatchLogsDestination - Undocumented member.

KinesisFirehoseDestination

data KinesisFirehoseDestination

An object that contains information about an event destination that sends data to Amazon Kinesis Data Firehose.

See: newKinesisFirehoseDestination smart constructor.

Constructors

KinesisFirehoseDestination' (Maybe Text) (Maybe Text)

newKinesisFirehoseDestination :: KinesisFirehoseDestination

Create a value of KinesisFirehoseDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamRoleArn:KinesisFirehoseDestination', kinesisFirehoseDestination_iamRoleArn - The Amazon Resource Name (ARN) of the Amazon Kinesis Data Firehose destination that you want to use in the event destination.

$sel:deliveryStreamArn:KinesisFirehoseDestination', kinesisFirehoseDestination_deliveryStreamArn - The Amazon Resource Name (ARN) of an IAM role that can write data to an Amazon Kinesis Data Firehose stream.

PlainTextMessageType

data PlainTextMessageType

An object that defines a message that contains unformatted text.

See: newPlainTextMessageType smart constructor.

Constructors

PlainTextMessageType' (Maybe Text) (Maybe Text) (Maybe Text)

newPlainTextMessageType :: PlainTextMessageType

Create a value of PlainTextMessageType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:PlainTextMessageType', plainTextMessageType_languageCode - The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

$sel:text:PlainTextMessageType', plainTextMessageType_text - The plain (not SSML-formatted) text to deliver to the recipient.

$sel:voiceId:PlainTextMessageType', plainTextMessageType_voiceId - The name of the voice that you want to use to deliver the message. For a complete list of supported voices, see the Amazon Polly Developer Guide.

SSMLMessageType

data SSMLMessageType

An object that defines a message that contains SSML-formatted text.

See: newSSMLMessageType smart constructor.

Constructors

SSMLMessageType' (Maybe Text) (Maybe Text) (Maybe Text)

newSSMLMessageType :: SSMLMessageType

Create a value of SSMLMessageType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:SSMLMessageType', sSMLMessageType_languageCode - The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

$sel:text:SSMLMessageType', sSMLMessageType_text - The SSML-formatted text to deliver to the recipient.

$sel:voiceId:SSMLMessageType', sSMLMessageType_voiceId - The name of the voice that you want to use to deliver the message. For a complete list of supported voices, see the Amazon Polly Developer Guide.

SnsDestination

data SnsDestination

An object that contains information about an event destination that sends data to Amazon SNS.

See: newSnsDestination smart constructor.

Constructors

SnsDestination' (Maybe Text)

newSnsDestination :: SnsDestination

Create a value of SnsDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:topicArn:SnsDestination', snsDestination_topicArn - The Amazon Resource Name (ARN) of the Amazon SNS topic that you want to publish events to.

VoiceMessageContent

data VoiceMessageContent

An object that contains a voice message and information about the recipient that you want to send it to.

See: newVoiceMessageContent smart constructor.

Constructors

VoiceMessageContent' (Maybe CallInstructionsMessageType) (Maybe SSMLMessageType) (Maybe PlainTextMessageType)

newVoiceMessageContent :: VoiceMessageContent

Create a value of VoiceMessageContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:callInstructionsMessage:VoiceMessageContent', voiceMessageContent_callInstructionsMessage - Undocumented member.

$sel:sSMLMessage:VoiceMessageContent', voiceMessageContent_sSMLMessage - Undocumented member.

$sel:plainTextMessage:VoiceMessageContent', voiceMessageContent_plainTextMessage - Undocumented member.