Amazonka.PinpointSMSVoice.DeleteConfigurationSet

Description

Deletes an existing configuration set.

Creating a Request

data DeleteConfigurationSet

See: newDeleteConfigurationSet smart constructor.

Constructors

DeleteConfigurationSet'

Fields

configurationSetName :: Text

ConfigurationSetName

newDeleteConfigurationSet

Arguments

:: Text
$sel:configurationSetName:DeleteConfigurationSet'

-> DeleteConfigurationSet

Create a value of DeleteConfigurationSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetName:DeleteConfigurationSet', deleteConfigurationSet_configurationSetName - ConfigurationSetName

Request Lenses

deleteConfigurationSet_configurationSetName :: Lens' DeleteConfigurationSet Text

ConfigurationSetName

Destructuring the Response

data DeleteConfigurationSetResponse

An empty object that indicates that the configuration set was deleted successfully.

See: newDeleteConfigurationSetResponse smart constructor.

Constructors

DeleteConfigurationSetResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteConfigurationSetResponse

Arguments

:: Int
$sel:httpStatus:DeleteConfigurationSetResponse'

-> DeleteConfigurationSetResponse

Create a value of DeleteConfigurationSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteConfigurationSetResponse', deleteConfigurationSetResponse_httpStatus - The response's http status code.

Response Lenses

deleteConfigurationSetResponse_httpStatus :: Lens' DeleteConfigurationSetResponse Int

The response's http status code.