Amazonka.PinpointSMSVoice.Lens

Description

Operations

UpdateConfigurationSetEventDestination

updateConfigurationSetEventDestination_eventDestination :: Lens' UpdateConfigurationSetEventDestination (Maybe EventDestinationDefinition)

Undocumented member.

updateConfigurationSetEventDestination_eventDestinationName :: Lens' UpdateConfigurationSetEventDestination Text

EventDestinationName

updateConfigurationSetEventDestination_configurationSetName :: Lens' UpdateConfigurationSetEventDestination Text

ConfigurationSetName

updateConfigurationSetEventDestinationResponse_httpStatus :: Lens' UpdateConfigurationSetEventDestinationResponse Int

The response's http status code.

DeleteConfigurationSetEventDestination

deleteConfigurationSetEventDestination_eventDestinationName :: Lens' DeleteConfigurationSetEventDestination Text

EventDestinationName

deleteConfigurationSetEventDestination_configurationSetName :: Lens' DeleteConfigurationSetEventDestination Text

ConfigurationSetName

deleteConfigurationSetEventDestinationResponse_httpStatus :: Lens' DeleteConfigurationSetEventDestinationResponse Int

The response's http status code.

DeleteConfigurationSet

deleteConfigurationSet_configurationSetName :: Lens' DeleteConfigurationSet Text

ConfigurationSetName

deleteConfigurationSetResponse_httpStatus :: Lens' DeleteConfigurationSetResponse Int

The response's http status code.

SendVoiceMessage

sendVoiceMessage_configurationSetName :: Lens' SendVoiceMessage (Maybe Text)

The name of the configuration set that you want to use to send the message.

sendVoiceMessage_callerId :: Lens' SendVoiceMessage (Maybe Text)

The phone number that appears on recipients' devices when they receive the message.

sendVoiceMessage_originationPhoneNumber :: Lens' SendVoiceMessage (Maybe Text)

The phone number that Amazon Pinpoint should use to send the voice message. This isn't necessarily the phone number that appears on recipients' devices when they receive the message, because you can specify a CallerId parameter in the request.

sendVoiceMessage_content :: Lens' SendVoiceMessage (Maybe VoiceMessageContent)

Undocumented member.

sendVoiceMessage_destinationPhoneNumber :: Lens' SendVoiceMessage (Maybe Text)

The phone number that you want to send the voice message to.

sendVoiceMessageResponse_messageId :: Lens' SendVoiceMessageResponse (Maybe Text)

A unique identifier for the voice message.

sendVoiceMessageResponse_httpStatus :: Lens' SendVoiceMessageResponse Int

The response's http status code.

GetConfigurationSetEventDestinations

getConfigurationSetEventDestinations_configurationSetName :: Lens' GetConfigurationSetEventDestinations Text

ConfigurationSetName

getConfigurationSetEventDestinationsResponse_eventDestinations :: Lens' GetConfigurationSetEventDestinationsResponse (Maybe [EventDestination])

Undocumented member.

getConfigurationSetEventDestinationsResponse_httpStatus :: Lens' GetConfigurationSetEventDestinationsResponse Int

The response's http status code.

CreateConfigurationSetEventDestination

createConfigurationSetEventDestination_eventDestination :: Lens' CreateConfigurationSetEventDestination (Maybe EventDestinationDefinition)

Undocumented member.

createConfigurationSetEventDestination_eventDestinationName :: Lens' CreateConfigurationSetEventDestination (Maybe Text)

A name that identifies the event destination.

createConfigurationSetEventDestination_configurationSetName :: Lens' CreateConfigurationSetEventDestination Text

ConfigurationSetName

createConfigurationSetEventDestinationResponse_httpStatus :: Lens' CreateConfigurationSetEventDestinationResponse Int

The response's http status code.

CreateConfigurationSet

createConfigurationSet_configurationSetName :: Lens' CreateConfigurationSet (Maybe Text)

The name that you want to give the configuration set.

createConfigurationSetResponse_httpStatus :: Lens' CreateConfigurationSetResponse Int

The response's http status code.

Types

CallInstructionsMessageType

callInstructionsMessageType_text :: Lens' CallInstructionsMessageType (Maybe Text)

The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

CloudWatchLogsDestination

cloudWatchLogsDestination_iamRoleArn :: Lens' CloudWatchLogsDestination (Maybe Text)

The Amazon Resource Name (ARN) of an Amazon Identity and Access Management (IAM) role that is able to write event data to an Amazon CloudWatch destination.

cloudWatchLogsDestination_logGroupArn :: Lens' CloudWatchLogsDestination (Maybe Text)

The name of the Amazon CloudWatch Log Group that you want to record events in.

EventDestination

eventDestination_matchingEventTypes :: Lens' EventDestination (Maybe [EventType])

Undocumented member.

eventDestination_enabled :: Lens' EventDestination (Maybe Bool)

Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

eventDestination_kinesisFirehoseDestination :: Lens' EventDestination (Maybe KinesisFirehoseDestination)

Undocumented member.

eventDestination_name :: Lens' EventDestination (Maybe Text)

A name that identifies the event destination configuration.

eventDestination_snsDestination :: Lens' EventDestination (Maybe SnsDestination)

Undocumented member.

eventDestination_cloudWatchLogsDestination :: Lens' EventDestination (Maybe CloudWatchLogsDestination)

Undocumented member.

EventDestinationDefinition

eventDestinationDefinition_matchingEventTypes :: Lens' EventDestinationDefinition (Maybe [EventType])

Undocumented member.

eventDestinationDefinition_enabled :: Lens' EventDestinationDefinition (Maybe Bool)

Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

eventDestinationDefinition_kinesisFirehoseDestination :: Lens' EventDestinationDefinition (Maybe KinesisFirehoseDestination)

Undocumented member.

eventDestinationDefinition_snsDestination :: Lens' EventDestinationDefinition (Maybe SnsDestination)

Undocumented member.

eventDestinationDefinition_cloudWatchLogsDestination :: Lens' EventDestinationDefinition (Maybe CloudWatchLogsDestination)

Undocumented member.

KinesisFirehoseDestination

kinesisFirehoseDestination_iamRoleArn :: Lens' KinesisFirehoseDestination (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis Data Firehose destination that you want to use in the event destination.

kinesisFirehoseDestination_deliveryStreamArn :: Lens' KinesisFirehoseDestination (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role that can write data to an Amazon Kinesis Data Firehose stream.

PlainTextMessageType

plainTextMessageType_languageCode :: Lens' PlainTextMessageType (Maybe Text)

The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

plainTextMessageType_text :: Lens' PlainTextMessageType (Maybe Text)

The plain (not SSML-formatted) text to deliver to the recipient.

plainTextMessageType_voiceId :: Lens' PlainTextMessageType (Maybe Text)

The name of the voice that you want to use to deliver the message. For a complete list of supported voices, see the Amazon Polly Developer Guide.

SSMLMessageType

sSMLMessageType_languageCode :: Lens' SSMLMessageType (Maybe Text)

The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

sSMLMessageType_text :: Lens' SSMLMessageType (Maybe Text)

The SSML-formatted text to deliver to the recipient.

sSMLMessageType_voiceId :: Lens' SSMLMessageType (Maybe Text)

The name of the voice that you want to use to deliver the message. For a complete list of supported voices, see the Amazon Polly Developer Guide.

SnsDestination

snsDestination_topicArn :: Lens' SnsDestination (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic that you want to publish events to.

VoiceMessageContent

voiceMessageContent_callInstructionsMessage :: Lens' VoiceMessageContent (Maybe CallInstructionsMessageType)

Undocumented member.

voiceMessageContent_sSMLMessage :: Lens' VoiceMessageContent (Maybe SSMLMessageType)

Undocumented member.

voiceMessageContent_plainTextMessage :: Lens' VoiceMessageContent (Maybe PlainTextMessageType)

Undocumented member.