Amazonka.PinpointSMSVoice.Types

Description

Service Configuration

defaultService :: Service

API version 2018-09-05 of the Amazon Pinpoint SMS and Voice Service SDK configuration.

Errors

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource you attempted to access doesn't exist.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

You've issued too many requests to the resource. Wait a few minutes, and then try again.

_InternalServiceErrorException :: AsError a => Getting (First ServiceError) a ServiceError

The API encountered an unexpected error and couldn't complete the request. You might be able to successfully issue the request again in the future.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The input you provided is invalid.

_AlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The resource specified in your request already exists.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

There are too many instances of the specified resource type.

EventType

newtype EventType

The types of events that are sent to the event destination.

Constructors

EventType'

Fields

fromEventType :: Text

Bundled Patterns

pattern EventType_ANSWERED :: EventType

pattern EventType_BUSY :: EventType

pattern EventType_COMPLETED_CALL :: EventType

pattern EventType_FAILED :: EventType

pattern EventType_INITIATED_CALL :: EventType

pattern EventType_NO_ANSWER :: EventType

pattern EventType_RINGING :: EventType

CallInstructionsMessageType

data CallInstructionsMessageType

An object that defines a message that contains text formatted using Amazon Pinpoint Voice Instructions markup.

See: newCallInstructionsMessageType smart constructor.

Constructors

CallInstructionsMessageType'

Fields

text :: Maybe Text

The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

newCallInstructionsMessageType :: CallInstructionsMessageType

Create a value of CallInstructionsMessageType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:text:CallInstructionsMessageType', callInstructionsMessageType_text - The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

callInstructionsMessageType_text :: Lens' CallInstructionsMessageType (Maybe Text)

The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

CloudWatchLogsDestination

data CloudWatchLogsDestination

An object that contains information about an event destination that sends data to Amazon CloudWatch Logs.

See: newCloudWatchLogsDestination smart constructor.

Constructors

CloudWatchLogsDestination'

Fields

iamRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of an Amazon Identity and Access Management (IAM) role that is able to write event data to an Amazon CloudWatch destination.

logGroupArn :: Maybe Text

The name of the Amazon CloudWatch Log Group that you want to record events in.

newCloudWatchLogsDestination :: CloudWatchLogsDestination

Create a value of CloudWatchLogsDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamRoleArn:CloudWatchLogsDestination', cloudWatchLogsDestination_iamRoleArn - The Amazon Resource Name (ARN) of an Amazon Identity and Access Management (IAM) role that is able to write event data to an Amazon CloudWatch destination.

$sel:logGroupArn:CloudWatchLogsDestination', cloudWatchLogsDestination_logGroupArn - The name of the Amazon CloudWatch Log Group that you want to record events in.

cloudWatchLogsDestination_iamRoleArn :: Lens' CloudWatchLogsDestination (Maybe Text)

The Amazon Resource Name (ARN) of an Amazon Identity and Access Management (IAM) role that is able to write event data to an Amazon CloudWatch destination.

cloudWatchLogsDestination_logGroupArn :: Lens' CloudWatchLogsDestination (Maybe Text)

The name of the Amazon CloudWatch Log Group that you want to record events in.

EventDestination

data EventDestination

An object that defines an event destination.

See: newEventDestination smart constructor.

Constructors

EventDestination'

Fields

matchingEventTypes :: Maybe [EventType]

enabled :: Maybe Bool

Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

kinesisFirehoseDestination :: Maybe KinesisFirehoseDestination

name :: Maybe Text

A name that identifies the event destination configuration.

snsDestination :: Maybe SnsDestination

cloudWatchLogsDestination :: Maybe CloudWatchLogsDestination

newEventDestination :: EventDestination

Create a value of EventDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:matchingEventTypes:EventDestination', eventDestination_matchingEventTypes - Undocumented member.

$sel:enabled:EventDestination', eventDestination_enabled - Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

$sel:kinesisFirehoseDestination:EventDestination', eventDestination_kinesisFirehoseDestination - Undocumented member.

$sel:name:EventDestination', eventDestination_name - A name that identifies the event destination configuration.

$sel:snsDestination:EventDestination', eventDestination_snsDestination - Undocumented member.

$sel:cloudWatchLogsDestination:EventDestination', eventDestination_cloudWatchLogsDestination - Undocumented member.

eventDestination_matchingEventTypes :: Lens' EventDestination (Maybe [EventType])

Undocumented member.

eventDestination_enabled :: Lens' EventDestination (Maybe Bool)

Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

eventDestination_kinesisFirehoseDestination :: Lens' EventDestination (Maybe KinesisFirehoseDestination)

Undocumented member.

eventDestination_name :: Lens' EventDestination (Maybe Text)

A name that identifies the event destination configuration.

eventDestination_snsDestination :: Lens' EventDestination (Maybe SnsDestination)

Undocumented member.

eventDestination_cloudWatchLogsDestination :: Lens' EventDestination (Maybe CloudWatchLogsDestination)

Undocumented member.

EventDestinationDefinition

data EventDestinationDefinition

An object that defines a single event destination.

See: newEventDestinationDefinition smart constructor.

Constructors

EventDestinationDefinition'

Fields

matchingEventTypes :: Maybe [EventType]

enabled :: Maybe Bool

Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

kinesisFirehoseDestination :: Maybe KinesisFirehoseDestination

snsDestination :: Maybe SnsDestination

cloudWatchLogsDestination :: Maybe CloudWatchLogsDestination

newEventDestinationDefinition :: EventDestinationDefinition

Create a value of EventDestinationDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:matchingEventTypes:EventDestinationDefinition', eventDestinationDefinition_matchingEventTypes - Undocumented member.

$sel:enabled:EventDestinationDefinition', eventDestinationDefinition_enabled - Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

$sel:kinesisFirehoseDestination:EventDestinationDefinition', eventDestinationDefinition_kinesisFirehoseDestination - Undocumented member.

$sel:snsDestination:EventDestinationDefinition', eventDestinationDefinition_snsDestination - Undocumented member.

$sel:cloudWatchLogsDestination:EventDestinationDefinition', eventDestinationDefinition_cloudWatchLogsDestination - Undocumented member.

eventDestinationDefinition_matchingEventTypes :: Lens' EventDestinationDefinition (Maybe [EventType])

Undocumented member.

eventDestinationDefinition_enabled :: Lens' EventDestinationDefinition (Maybe Bool)

Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

eventDestinationDefinition_kinesisFirehoseDestination :: Lens' EventDestinationDefinition (Maybe KinesisFirehoseDestination)

Undocumented member.

eventDestinationDefinition_snsDestination :: Lens' EventDestinationDefinition (Maybe SnsDestination)

Undocumented member.

eventDestinationDefinition_cloudWatchLogsDestination :: Lens' EventDestinationDefinition (Maybe CloudWatchLogsDestination)

Undocumented member.

KinesisFirehoseDestination

data KinesisFirehoseDestination

An object that contains information about an event destination that sends data to Amazon Kinesis Data Firehose.

See: newKinesisFirehoseDestination smart constructor.

Constructors

KinesisFirehoseDestination'

Fields

iamRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon Kinesis Data Firehose destination that you want to use in the event destination.

deliveryStreamArn :: Maybe Text

The Amazon Resource Name (ARN) of an IAM role that can write data to an Amazon Kinesis Data Firehose stream.

newKinesisFirehoseDestination :: KinesisFirehoseDestination

Create a value of KinesisFirehoseDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamRoleArn:KinesisFirehoseDestination', kinesisFirehoseDestination_iamRoleArn - The Amazon Resource Name (ARN) of the Amazon Kinesis Data Firehose destination that you want to use in the event destination.

$sel:deliveryStreamArn:KinesisFirehoseDestination', kinesisFirehoseDestination_deliveryStreamArn - The Amazon Resource Name (ARN) of an IAM role that can write data to an Amazon Kinesis Data Firehose stream.

kinesisFirehoseDestination_iamRoleArn :: Lens' KinesisFirehoseDestination (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Kinesis Data Firehose destination that you want to use in the event destination.

kinesisFirehoseDestination_deliveryStreamArn :: Lens' KinesisFirehoseDestination (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role that can write data to an Amazon Kinesis Data Firehose stream.

PlainTextMessageType

data PlainTextMessageType

An object that defines a message that contains unformatted text.

See: newPlainTextMessageType smart constructor.

Constructors

PlainTextMessageType'

Fields

languageCode :: Maybe Text

The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

text :: Maybe Text

The plain (not SSML-formatted) text to deliver to the recipient.

voiceId :: Maybe Text

The name of the voice that you want to use to deliver the message. For a complete list of supported voices, see the Amazon Polly Developer Guide.

newPlainTextMessageType :: PlainTextMessageType

Create a value of PlainTextMessageType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:PlainTextMessageType', plainTextMessageType_languageCode - The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

$sel:text:PlainTextMessageType', plainTextMessageType_text - The plain (not SSML-formatted) text to deliver to the recipient.

$sel:voiceId:PlainTextMessageType', plainTextMessageType_voiceId - The name of the voice that you want to use to deliver the message. For a complete list of supported voices, see the Amazon Polly Developer Guide.

plainTextMessageType_languageCode :: Lens' PlainTextMessageType (Maybe Text)

The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

plainTextMessageType_text :: Lens' PlainTextMessageType (Maybe Text)

The plain (not SSML-formatted) text to deliver to the recipient.

plainTextMessageType_voiceId :: Lens' PlainTextMessageType (Maybe Text)

The name of the voice that you want to use to deliver the message. For a complete list of supported voices, see the Amazon Polly Developer Guide.

SSMLMessageType

data SSMLMessageType

An object that defines a message that contains SSML-formatted text.

See: newSSMLMessageType smart constructor.

Constructors

SSMLMessageType'

Fields

languageCode :: Maybe Text

The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

text :: Maybe Text

The SSML-formatted text to deliver to the recipient.

voiceId :: Maybe Text

The name of the voice that you want to use to deliver the message. For a complete list of supported voices, see the Amazon Polly Developer Guide.

newSSMLMessageType :: SSMLMessageType

Create a value of SSMLMessageType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:SSMLMessageType', sSMLMessageType_languageCode - The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

$sel:text:SSMLMessageType', sSMLMessageType_text - The SSML-formatted text to deliver to the recipient.

$sel:voiceId:SSMLMessageType', sSMLMessageType_voiceId - The name of the voice that you want to use to deliver the message. For a complete list of supported voices, see the Amazon Polly Developer Guide.

sSMLMessageType_languageCode :: Lens' SSMLMessageType (Maybe Text)

The language to use when delivering the message. For a complete list of supported languages, see the Amazon Polly Developer Guide.

sSMLMessageType_text :: Lens' SSMLMessageType (Maybe Text)

The SSML-formatted text to deliver to the recipient.

sSMLMessageType_voiceId :: Lens' SSMLMessageType (Maybe Text)

The name of the voice that you want to use to deliver the message. For a complete list of supported voices, see the Amazon Polly Developer Guide.

SnsDestination

data SnsDestination

An object that contains information about an event destination that sends data to Amazon SNS.

See: newSnsDestination smart constructor.

Constructors

SnsDestination'

Fields

topicArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon SNS topic that you want to publish events to.

newSnsDestination :: SnsDestination

Create a value of SnsDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:topicArn:SnsDestination', snsDestination_topicArn - The Amazon Resource Name (ARN) of the Amazon SNS topic that you want to publish events to.

snsDestination_topicArn :: Lens' SnsDestination (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon SNS topic that you want to publish events to.

VoiceMessageContent

data VoiceMessageContent

An object that contains a voice message and information about the recipient that you want to send it to.

See: newVoiceMessageContent smart constructor.

Constructors

VoiceMessageContent'

Fields

callInstructionsMessage :: Maybe CallInstructionsMessageType

sSMLMessage :: Maybe SSMLMessageType

plainTextMessage :: Maybe PlainTextMessageType

newVoiceMessageContent :: VoiceMessageContent

Create a value of VoiceMessageContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:callInstructionsMessage:VoiceMessageContent', voiceMessageContent_callInstructionsMessage - Undocumented member.

$sel:sSMLMessage:VoiceMessageContent', voiceMessageContent_sSMLMessage - Undocumented member.

$sel:plainTextMessage:VoiceMessageContent', voiceMessageContent_plainTextMessage - Undocumented member.

voiceMessageContent_callInstructionsMessage :: Lens' VoiceMessageContent (Maybe CallInstructionsMessageType)

Undocumented member.

voiceMessageContent_sSMLMessage :: Lens' VoiceMessageContent (Maybe SSMLMessageType)

Undocumented member.

voiceMessageContent_plainTextMessage :: Lens' VoiceMessageContent (Maybe PlainTextMessageType)

Undocumented member.