Amazonka.PinpointSMSVoice.Types.EventDestinationDefinition

Description

Documentation

data EventDestinationDefinition

An object that defines a single event destination.

See: newEventDestinationDefinition smart constructor.

Constructors

EventDestinationDefinition'

Fields

matchingEventTypes :: Maybe [EventType]

enabled :: Maybe Bool

Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

kinesisFirehoseDestination :: Maybe KinesisFirehoseDestination

snsDestination :: Maybe SnsDestination

cloudWatchLogsDestination :: Maybe CloudWatchLogsDestination

newEventDestinationDefinition :: EventDestinationDefinition

Create a value of EventDestinationDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:matchingEventTypes:EventDestinationDefinition', eventDestinationDefinition_matchingEventTypes - Undocumented member.

$sel:enabled:EventDestinationDefinition', eventDestinationDefinition_enabled - Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

$sel:kinesisFirehoseDestination:EventDestinationDefinition', eventDestinationDefinition_kinesisFirehoseDestination - Undocumented member.

$sel:snsDestination:EventDestinationDefinition', eventDestinationDefinition_snsDestination - Undocumented member.

$sel:cloudWatchLogsDestination:EventDestinationDefinition', eventDestinationDefinition_cloudWatchLogsDestination - Undocumented member.

eventDestinationDefinition_matchingEventTypes :: Lens' EventDestinationDefinition (Maybe [EventType])

Undocumented member.

eventDestinationDefinition_enabled :: Lens' EventDestinationDefinition (Maybe Bool)

Indicates whether or not the event destination is enabled. If the event destination is enabled, then Amazon Pinpoint sends response data to the specified event destination.

eventDestinationDefinition_kinesisFirehoseDestination :: Lens' EventDestinationDefinition (Maybe KinesisFirehoseDestination)

Undocumented member.

eventDestinationDefinition_snsDestination :: Lens' EventDestinationDefinition (Maybe SnsDestination)

Undocumented member.

eventDestinationDefinition_cloudWatchLogsDestination :: Lens' EventDestinationDefinition (Maybe CloudWatchLogsDestination)

Undocumented member.