Amazonka.Polly

Description

Derived from API version 2016-06-10 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Polly is a web service that makes it easy to synthesize speech from text.

The Amazon Polly service provides API operations for synthesizing high-quality speech from plain text and Speech Synthesis Markup Language (SSML), along with managing pronunciations lexicons that enable you to get the best results for your application domain.

Service Configuration

defaultService :: Service

API version 2016-06-10 of the Amazon Polly SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Polly.

InvalidSnsTopicArnException

_InvalidSnsTopicArnException :: AsError a => Getting (First ServiceError) a ServiceError

The provided SNS topic ARN is invalid. Please provide a valid SNS topic ARN and try again.

UnsupportedPlsLanguageException

_UnsupportedPlsLanguageException :: AsError a => Getting (First ServiceError) a ServiceError

The language specified in the lexicon is unsupported. For a list of supported languages, see Lexicon Attributes.

InvalidSsmlException

_InvalidSsmlException :: AsError a => Getting (First ServiceError) a ServiceError

The SSML you provided is invalid. Verify the SSML syntax, spelling of tags and values, and then try again.

InvalidSampleRateException

_InvalidSampleRateException :: AsError a => Getting (First ServiceError) a ServiceError

The specified sample rate is not valid.

EngineNotSupportedException

_EngineNotSupportedException :: AsError a => Getting (First ServiceError) a ServiceError

This engine is not compatible with the voice that you have designated. Choose a new voice that is compatible with the engine or change the engine and restart the operation.

MaxLexiconsNumberExceededException

_MaxLexiconsNumberExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of lexicons would be exceeded by this operation.

TextLengthExceededException

_TextLengthExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The value of the "Text" parameter is longer than the accepted limits. For the SynthesizeSpeech API, the limit for input text is a maximum of 6000 characters total, of which no more than 3000 can be billed characters. For the StartSpeechSynthesisTask API, the maximum is 200,000 characters, of which no more than 100,000 can be billed characters. SSML tags are not counted as billed characters.

MaxLexemeLengthExceededException

_MaxLexemeLengthExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum size of the lexeme would be exceeded by this operation.

InvalidTaskIdException

_InvalidTaskIdException :: AsError a => Getting (First ServiceError) a ServiceError

The provided Task ID is not valid. Please provide a valid Task ID and try again.

InvalidLexiconException

_InvalidLexiconException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Polly can't find the specified lexicon. Verify that the lexicon's name is spelled correctly, and then try again.

ServiceFailureException

_ServiceFailureException :: AsError a => Getting (First ServiceError) a ServiceError

An unknown condition has caused a service failure.

UnsupportedPlsAlphabetException

_UnsupportedPlsAlphabetException :: AsError a => Getting (First ServiceError) a ServiceError

The alphabet specified by the lexicon is not a supported alphabet. Valid values are x-sampa and ipa.

InvalidNextTokenException

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The NextToken is invalid. Verify that it's spelled correctly, and then try again.

MarksNotSupportedForFormatException

_MarksNotSupportedForFormatException :: AsError a => Getting (First ServiceError) a ServiceError

Speech marks are not supported for the OutputFormat selected. Speech marks are only available for content in json format.

SynthesisTaskNotFoundException

_SynthesisTaskNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The Speech Synthesis task with requested Task ID cannot be found.

SsmlMarksNotSupportedForTextTypeException

_SsmlMarksNotSupportedForTextTypeException :: AsError a => Getting (First ServiceError) a ServiceError

SSML speech marks are not supported for plain text-type input.

InvalidS3BucketException

_InvalidS3BucketException :: AsError a => Getting (First ServiceError) a ServiceError

The provided Amazon S3 bucket name is invalid. Please check your input with S3 bucket naming requirements and try again.

LexiconSizeExceededException

_LexiconSizeExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum size of the specified lexicon would be exceeded by this operation.

LanguageNotSupportedException

_LanguageNotSupportedException :: AsError a => Getting (First ServiceError) a ServiceError

The language specified is not currently supported by Amazon Polly in this capacity.

LexiconNotFoundException

_LexiconNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Polly can't find the specified lexicon. This could be caused by a lexicon that is missing, its name is misspelled or specifying a lexicon that is in a different region.

Verify that the lexicon exists, is in the region (see ListLexicons) and that you spelled its name is spelled correctly. Then try again.

InvalidS3KeyException

_InvalidS3KeyException :: AsError a => Getting (First ServiceError) a ServiceError

The provided Amazon S3 key prefix is invalid. Please provide a valid S3 object key name.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetLexicon

data GetLexicon

See: newGetLexicon smart constructor.

Constructors

GetLexicon' Text

newGetLexicon

Arguments

:: Text
$sel:name:GetLexicon'

-> GetLexicon

Create a value of GetLexicon with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetLexicon', getLexicon_name - Name of the lexicon.

data GetLexiconResponse

See: newGetLexiconResponse smart constructor.

Constructors

GetLexiconResponse' (Maybe LexiconAttributes) (Maybe Lexicon) Int

newGetLexiconResponse

Arguments

:: Int
$sel:httpStatus:GetLexiconResponse'

-> GetLexiconResponse

Create a value of GetLexiconResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lexiconAttributes:GetLexiconResponse', getLexiconResponse_lexiconAttributes - Metadata of the lexicon, including phonetic alphabetic used, language code, lexicon ARN, number of lexemes defined in the lexicon, and size of lexicon in bytes.

$sel:lexicon:GetLexiconResponse', getLexiconResponse_lexicon - Lexicon object that provides name and the string content of the lexicon.

$sel:httpStatus:GetLexiconResponse', getLexiconResponse_httpStatus - The response's http status code.

GetSpeechSynthesisTask

data GetSpeechSynthesisTask

See: newGetSpeechSynthesisTask smart constructor.

Constructors

GetSpeechSynthesisTask' Text

newGetSpeechSynthesisTask

Arguments

:: Text
$sel:taskId:GetSpeechSynthesisTask'

-> GetSpeechSynthesisTask

Create a value of GetSpeechSynthesisTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:GetSpeechSynthesisTask', getSpeechSynthesisTask_taskId - The Amazon Polly generated identifier for a speech synthesis task.

data GetSpeechSynthesisTaskResponse

See: newGetSpeechSynthesisTaskResponse smart constructor.

Constructors

GetSpeechSynthesisTaskResponse' (Maybe SynthesisTask) Int

newGetSpeechSynthesisTaskResponse

Arguments

:: Int
$sel:httpStatus:GetSpeechSynthesisTaskResponse'

-> GetSpeechSynthesisTaskResponse

Create a value of GetSpeechSynthesisTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:synthesisTask:GetSpeechSynthesisTaskResponse', getSpeechSynthesisTaskResponse_synthesisTask - SynthesisTask object that provides information from the requested task, including output format, creation time, task status, and so on.

$sel:httpStatus:GetSpeechSynthesisTaskResponse', getSpeechSynthesisTaskResponse_httpStatus - The response's http status code.

DescribeVoices (Paginated)

data DescribeVoices

See: newDescribeVoices smart constructor.

Constructors

DescribeVoices' (Maybe LanguageCode) (Maybe Engine) (Maybe Text) (Maybe Bool)

newDescribeVoices :: DescribeVoices

Create a value of DescribeVoices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:DescribeVoices', describeVoices_languageCode - The language identification tag (ISO 639 code for the language name-ISO 3166 country code) for filtering the list of voices returned. If you don't specify this optional parameter, all available voices are returned.

$sel:engine:DescribeVoices', describeVoices_engine - Specifies the engine (standard or neural) used by Amazon Polly when processing input text for speech synthesis.

$sel:nextToken:DescribeVoices', describeVoices_nextToken - An opaque pagination token returned from the previous DescribeVoices operation. If present, this indicates where to continue the listing.

$sel:includeAdditionalLanguageCodes:DescribeVoices', describeVoices_includeAdditionalLanguageCodes - Boolean value indicating whether to return any bilingual voices that use the specified language as an additional language. For instance, if you request all languages that use US English (es-US), and there is an Italian voice that speaks both Italian (it-IT) and US English, that voice will be included if you specify yes but not if you specify no.

data DescribeVoicesResponse

See: newDescribeVoicesResponse smart constructor.

Constructors

DescribeVoicesResponse' (Maybe Text) (Maybe [Voice]) Int

newDescribeVoicesResponse

Arguments

:: Int
$sel:httpStatus:DescribeVoicesResponse'

-> DescribeVoicesResponse

Create a value of DescribeVoicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeVoices', describeVoicesResponse_nextToken - The pagination token to use in the next request to continue the listing of voices. NextToken is returned only if the response is truncated.

$sel:voices:DescribeVoicesResponse', describeVoicesResponse_voices - A list of voices with their properties.

$sel:httpStatus:DescribeVoicesResponse', describeVoicesResponse_httpStatus - The response's http status code.

ListLexicons (Paginated)

data ListLexicons

See: newListLexicons smart constructor.

Constructors

ListLexicons' (Maybe Text)

newListLexicons :: ListLexicons

Create a value of ListLexicons with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLexicons', listLexicons_nextToken - An opaque pagination token returned from previous ListLexicons operation. If present, indicates where to continue the list of lexicons.

data ListLexiconsResponse

See: newListLexiconsResponse smart constructor.

Constructors

ListLexiconsResponse' (Maybe [LexiconDescription]) (Maybe Text) Int

newListLexiconsResponse

Arguments

:: Int
$sel:httpStatus:ListLexiconsResponse'

-> ListLexiconsResponse

Create a value of ListLexiconsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lexicons:ListLexiconsResponse', listLexiconsResponse_lexicons - A list of lexicon names and attributes.

$sel:nextToken:ListLexicons', listLexiconsResponse_nextToken - The pagination token to use in the next request to continue the listing of lexicons. NextToken is returned only if the response is truncated.

$sel:httpStatus:ListLexiconsResponse', listLexiconsResponse_httpStatus - The response's http status code.

SynthesizeSpeech

data SynthesizeSpeech

See: newSynthesizeSpeech smart constructor.

Constructors

SynthesizeSpeech' (Maybe LanguageCode) (Maybe Engine) (Maybe [SpeechMarkType]) (Maybe Text) (Maybe TextType) (Maybe [Text]) OutputFormat Text VoiceId

newSynthesizeSpeech

Arguments

:: OutputFormat
$sel:outputFormat:SynthesizeSpeech'

-> Text
$sel:text:SynthesizeSpeech'

-> VoiceId
$sel:voiceId:SynthesizeSpeech'

-> SynthesizeSpeech

Create a value of SynthesizeSpeech with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:SynthesizeSpeech', synthesizeSpeech_languageCode - Optional language code for the Synthesize Speech request. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly uses the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

$sel:engine:SynthesizeSpeech', synthesizeSpeech_engine - Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. For information on Amazon Polly voices and which voices are available in standard-only, NTTS-only, and both standard and NTTS formats, see Available Voices.

NTTS-only voices

When using NTTS-only voices such as Kevin (en-US), this parameter is required and must be set to neural. If the engine is not specified, or is set to standard, this will result in an error.

Type: String

Valid Values: standard | neural

Required: Yes

Standard voices

For standard voices, this is not required; the engine parameter defaults to standard. If the engine is not specified, or is set to standard and an NTTS-only voice is selected, this will result in an error.

$sel:speechMarkTypes:SynthesizeSpeech', synthesizeSpeech_speechMarkTypes - The type of speech marks returned for the input text.

$sel:sampleRate:SynthesizeSpeech', synthesizeSpeech_sampleRate - The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

$sel:textType:SynthesizeSpeech', synthesizeSpeech_textType - Specifies whether the input text is plain text or SSML. The default value is plain text. For more information, see Using SSML.

$sel:lexiconNames:SynthesizeSpeech', synthesizeSpeech_lexiconNames - List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice. For information about storing lexicons, see PutLexicon.

$sel:outputFormat:SynthesizeSpeech', synthesizeSpeech_outputFormat - The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

When pcm is used, the content returned is audio/pcm in a signed 16-bit, 1 channel (mono), little-endian format.

$sel:text:SynthesizeSpeech', synthesizeSpeech_text - Input text to synthesize. If you specify ssml as the TextType, follow the SSML format for the input text.

$sel:voiceId:SynthesizeSpeech', synthesizeSpeech_voiceId - Voice ID to use for the synthesis. You can get a list of available voice IDs by calling the DescribeVoices operation.

data SynthesizeSpeechResponse

See: newSynthesizeSpeechResponse smart constructor.

Constructors

SynthesizeSpeechResponse' (Maybe Int) (Maybe Text) Int ResponseBody

newSynthesizeSpeechResponse

Arguments

:: Int
$sel:httpStatus:SynthesizeSpeechResponse'

-> ResponseBody
$sel:audioStream:SynthesizeSpeechResponse'

-> SynthesizeSpeechResponse

Create a value of SynthesizeSpeechResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharacters:SynthesizeSpeechResponse', synthesizeSpeechResponse_requestCharacters - Number of characters synthesized.

$sel:contentType:SynthesizeSpeechResponse', synthesizeSpeechResponse_contentType - Specifies the type audio stream. This should reflect the OutputFormat parameter in your request.

If you request mp3 as the OutputFormat, the ContentType returned is audio/mpeg.

If you request ogg_vorbis as the OutputFormat, the ContentType returned is audio/ogg.

If you request pcm as the OutputFormat, the ContentType returned is audio/pcm in a signed 16-bit, 1 channel (mono), little-endian format.

If you request json as the OutputFormat, the ContentType returned is audio/json.

$sel:httpStatus:SynthesizeSpeechResponse', synthesizeSpeechResponse_httpStatus - The response's http status code.

$sel:audioStream:SynthesizeSpeechResponse', synthesizeSpeechResponse_audioStream - Stream containing the synthesized speech.

ListSpeechSynthesisTasks (Paginated)

data ListSpeechSynthesisTasks

See: newListSpeechSynthesisTasks smart constructor.

Constructors

ListSpeechSynthesisTasks' (Maybe TaskStatus) (Maybe Text) (Maybe Natural)

newListSpeechSynthesisTasks :: ListSpeechSynthesisTasks

Create a value of ListSpeechSynthesisTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListSpeechSynthesisTasks', listSpeechSynthesisTasks_status - Status of the speech synthesis tasks returned in a List operation

$sel:nextToken:ListSpeechSynthesisTasks', listSpeechSynthesisTasks_nextToken - The pagination token to use in the next request to continue the listing of speech synthesis tasks.

$sel:maxResults:ListSpeechSynthesisTasks', listSpeechSynthesisTasks_maxResults - Maximum number of speech synthesis tasks returned in a List operation.

data ListSpeechSynthesisTasksResponse

See: newListSpeechSynthesisTasksResponse smart constructor.

Constructors

ListSpeechSynthesisTasksResponse' (Maybe Text) (Maybe [SynthesisTask]) Int

newListSpeechSynthesisTasksResponse

Arguments

:: Int
$sel:httpStatus:ListSpeechSynthesisTasksResponse'

-> ListSpeechSynthesisTasksResponse

Create a value of ListSpeechSynthesisTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSpeechSynthesisTasks', listSpeechSynthesisTasksResponse_nextToken - An opaque pagination token returned from the previous List operation in this request. If present, this indicates where to continue the listing.

$sel:synthesisTasks:ListSpeechSynthesisTasksResponse', listSpeechSynthesisTasksResponse_synthesisTasks - List of SynthesisTask objects that provides information from the specified task in the list request, including output format, creation time, task status, and so on.

$sel:httpStatus:ListSpeechSynthesisTasksResponse', listSpeechSynthesisTasksResponse_httpStatus - The response's http status code.

PutLexicon

data PutLexicon

See: newPutLexicon smart constructor.

Constructors

PutLexicon' Text (Sensitive Text)

newPutLexicon

Arguments

:: Text
$sel:name:PutLexicon'

-> Text
$sel:content:PutLexicon'

-> PutLexicon

Create a value of PutLexicon with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:PutLexicon', putLexicon_name - Name of the lexicon. The name must follow the regular express format [0-9A-Za-z]{1,20}. That is, the name is a case-sensitive alphanumeric string up to 20 characters long.

$sel:content:PutLexicon', putLexicon_content - Content of the PLS lexicon as string data.

data PutLexiconResponse

See: newPutLexiconResponse smart constructor.

Constructors

PutLexiconResponse' Int

newPutLexiconResponse

Arguments

:: Int
$sel:httpStatus:PutLexiconResponse'

-> PutLexiconResponse

Create a value of PutLexiconResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutLexiconResponse', putLexiconResponse_httpStatus - The response's http status code.

DeleteLexicon

data DeleteLexicon

See: newDeleteLexicon smart constructor.

Constructors

DeleteLexicon' Text

newDeleteLexicon

Arguments

:: Text
$sel:name:DeleteLexicon'

-> DeleteLexicon

Create a value of DeleteLexicon with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteLexicon', deleteLexicon_name - The name of the lexicon to delete. Must be an existing lexicon in the region.

data DeleteLexiconResponse

See: newDeleteLexiconResponse smart constructor.

Constructors

DeleteLexiconResponse' Int

newDeleteLexiconResponse

Arguments

:: Int
$sel:httpStatus:DeleteLexiconResponse'

-> DeleteLexiconResponse

Create a value of DeleteLexiconResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteLexiconResponse', deleteLexiconResponse_httpStatus - The response's http status code.

StartSpeechSynthesisTask

data StartSpeechSynthesisTask

See: newStartSpeechSynthesisTask smart constructor.

Constructors

StartSpeechSynthesisTask' (Maybe LanguageCode) (Maybe Text) (Maybe Text) (Maybe Engine) (Maybe [SpeechMarkType]) (Maybe Text) (Maybe TextType) (Maybe [Text]) OutputFormat Text Text VoiceId

newStartSpeechSynthesisTask

Arguments

:: OutputFormat
$sel:outputFormat:StartSpeechSynthesisTask'

-> Text
$sel:outputS3BucketName:StartSpeechSynthesisTask'

-> Text
$sel:text:StartSpeechSynthesisTask'

-> VoiceId
$sel:voiceId:StartSpeechSynthesisTask'

-> StartSpeechSynthesisTask

Create a value of StartSpeechSynthesisTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:StartSpeechSynthesisTask', startSpeechSynthesisTask_languageCode - Optional language code for the Speech Synthesis request. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly uses the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

$sel:snsTopicArn:StartSpeechSynthesisTask', startSpeechSynthesisTask_snsTopicArn - ARN for the SNS topic optionally used for providing status notification for a speech synthesis task.

$sel:outputS3KeyPrefix:StartSpeechSynthesisTask', startSpeechSynthesisTask_outputS3KeyPrefix - The Amazon S3 key prefix for the output speech file.

$sel:engine:StartSpeechSynthesisTask', startSpeechSynthesisTask_engine - Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. Using a voice that is not supported for the engine selected will result in an error.

$sel:speechMarkTypes:StartSpeechSynthesisTask', startSpeechSynthesisTask_speechMarkTypes - The type of speech marks returned for the input text.

$sel:sampleRate:StartSpeechSynthesisTask', startSpeechSynthesisTask_sampleRate - The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

$sel:textType:StartSpeechSynthesisTask', startSpeechSynthesisTask_textType - Specifies whether the input text is plain text or SSML. The default value is plain text.

$sel:lexiconNames:StartSpeechSynthesisTask', startSpeechSynthesisTask_lexiconNames - List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice.

$sel:outputFormat:StartSpeechSynthesisTask', startSpeechSynthesisTask_outputFormat - The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

$sel:outputS3BucketName:StartSpeechSynthesisTask', startSpeechSynthesisTask_outputS3BucketName - Amazon S3 bucket name to which the output file will be saved.

$sel:text:StartSpeechSynthesisTask', startSpeechSynthesisTask_text - The input text to synthesize. If you specify ssml as the TextType, follow the SSML format for the input text.

$sel:voiceId:StartSpeechSynthesisTask', startSpeechSynthesisTask_voiceId - Voice ID to use for the synthesis.

data StartSpeechSynthesisTaskResponse

See: newStartSpeechSynthesisTaskResponse smart constructor.

Constructors

StartSpeechSynthesisTaskResponse' (Maybe SynthesisTask) Int

newStartSpeechSynthesisTaskResponse

Arguments

:: Int
$sel:httpStatus:StartSpeechSynthesisTaskResponse'

-> StartSpeechSynthesisTaskResponse

Create a value of StartSpeechSynthesisTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:synthesisTask:StartSpeechSynthesisTaskResponse', startSpeechSynthesisTaskResponse_synthesisTask - SynthesisTask object that provides information and attributes about a newly submitted speech synthesis task.

$sel:httpStatus:StartSpeechSynthesisTaskResponse', startSpeechSynthesisTaskResponse_httpStatus - The response's http status code.

Types

Engine

newtype Engine

Constructors

Engine'

Fields

fromEngine :: Text

Bundled Patterns

pattern Engine_Neural :: Engine

pattern Engine_Standard :: Engine

Gender

newtype Gender

Constructors

Gender'

Fields

fromGender :: Text

Bundled Patterns

pattern Gender_Female :: Gender

pattern Gender_Male :: Gender

LanguageCode

newtype LanguageCode

Constructors

LanguageCode'

Fields

fromLanguageCode :: Text

Bundled Patterns

pattern LanguageCode_Arb :: LanguageCode

pattern LanguageCode_Cmn_CN :: LanguageCode

pattern LanguageCode_Cy_GB :: LanguageCode

pattern LanguageCode_Da_DK :: LanguageCode

pattern LanguageCode_De_DE :: LanguageCode

pattern LanguageCode_En_AU :: LanguageCode

pattern LanguageCode_En_GB :: LanguageCode

pattern LanguageCode_En_GB_WLS :: LanguageCode

pattern LanguageCode_En_IN :: LanguageCode

pattern LanguageCode_En_NZ :: LanguageCode

pattern LanguageCode_En_US :: LanguageCode

pattern LanguageCode_En_ZA :: LanguageCode

pattern LanguageCode_Es_ES :: LanguageCode

pattern LanguageCode_Es_MX :: LanguageCode

pattern LanguageCode_Es_US :: LanguageCode

pattern LanguageCode_Fr_CA :: LanguageCode

pattern LanguageCode_Fr_FR :: LanguageCode

pattern LanguageCode_Hi_IN :: LanguageCode

pattern LanguageCode_Is_IS :: LanguageCode

pattern LanguageCode_It_IT :: LanguageCode

pattern LanguageCode_Ja_JP :: LanguageCode

pattern LanguageCode_Ko_KR :: LanguageCode

pattern LanguageCode_Nb_NO :: LanguageCode

pattern LanguageCode_Nl_NL :: LanguageCode

pattern LanguageCode_Pl_PL :: LanguageCode

pattern LanguageCode_Pt_BR :: LanguageCode

pattern LanguageCode_Pt_PT :: LanguageCode

pattern LanguageCode_Ro_RO :: LanguageCode

pattern LanguageCode_Ru_RU :: LanguageCode

pattern LanguageCode_Sv_SE :: LanguageCode

pattern LanguageCode_Tr_TR :: LanguageCode

OutputFormat

newtype OutputFormat

Constructors

OutputFormat'

Fields

fromOutputFormat :: Text

Bundled Patterns

pattern OutputFormat_Json :: OutputFormat

pattern OutputFormat_Mp3 :: OutputFormat

pattern OutputFormat_Ogg_vorbis :: OutputFormat

pattern OutputFormat_Pcm :: OutputFormat

SpeechMarkType

newtype SpeechMarkType

Constructors

SpeechMarkType'

Fields

fromSpeechMarkType :: Text

Bundled Patterns

pattern SpeechMarkType_Sentence :: SpeechMarkType

pattern SpeechMarkType_Ssml :: SpeechMarkType

pattern SpeechMarkType_Viseme :: SpeechMarkType

pattern SpeechMarkType_Word :: SpeechMarkType

TaskStatus

newtype TaskStatus

Constructors

TaskStatus'

Fields

fromTaskStatus :: Text

Bundled Patterns

pattern TaskStatus_Completed :: TaskStatus

pattern TaskStatus_Failed :: TaskStatus

pattern TaskStatus_InProgress :: TaskStatus

pattern TaskStatus_Scheduled :: TaskStatus

TextType

newtype TextType

Constructors

TextType'

Fields

fromTextType :: Text

Bundled Patterns

pattern TextType_Ssml :: TextType

pattern TextType_Text :: TextType

VoiceId

newtype VoiceId

Constructors

VoiceId'

Fields

fromVoiceId :: Text

Bundled Patterns

pattern VoiceId_Aditi :: VoiceId

pattern VoiceId_Amy :: VoiceId

pattern VoiceId_Aria :: VoiceId

pattern VoiceId_Astrid :: VoiceId

pattern VoiceId_Ayanda :: VoiceId

pattern VoiceId_Bianca :: VoiceId

pattern VoiceId_Brian :: VoiceId

pattern VoiceId_Camila :: VoiceId

pattern VoiceId_Carla :: VoiceId

pattern VoiceId_Carmen :: VoiceId

pattern VoiceId_Celine :: VoiceId

pattern VoiceId_Chantal :: VoiceId

pattern VoiceId_Conchita :: VoiceId

pattern VoiceId_Cristiano :: VoiceId

pattern VoiceId_Dora :: VoiceId

pattern VoiceId_Emma :: VoiceId

pattern VoiceId_Enrique :: VoiceId

pattern VoiceId_Ewa :: VoiceId

pattern VoiceId_Filiz :: VoiceId

pattern VoiceId_Gabrielle :: VoiceId

pattern VoiceId_Geraint :: VoiceId

pattern VoiceId_Giorgio :: VoiceId

pattern VoiceId_Gwyneth :: VoiceId

pattern VoiceId_Hans :: VoiceId

pattern VoiceId_Ines :: VoiceId

pattern VoiceId_Ivy :: VoiceId

pattern VoiceId_Jacek :: VoiceId

pattern VoiceId_Jan :: VoiceId

pattern VoiceId_Joanna :: VoiceId

pattern VoiceId_Joey :: VoiceId

pattern VoiceId_Justin :: VoiceId

pattern VoiceId_Karl :: VoiceId

pattern VoiceId_Kendra :: VoiceId

pattern VoiceId_Kevin :: VoiceId

pattern VoiceId_Kimberly :: VoiceId

pattern VoiceId_Lea :: VoiceId

pattern VoiceId_Liv :: VoiceId

pattern VoiceId_Lotte :: VoiceId

pattern VoiceId_Lucia :: VoiceId

pattern VoiceId_Lupe :: VoiceId

pattern VoiceId_Mads :: VoiceId

pattern VoiceId_Maja :: VoiceId

pattern VoiceId_Marlene :: VoiceId

pattern VoiceId_Mathieu :: VoiceId

pattern VoiceId_Matthew :: VoiceId

pattern VoiceId_Maxim :: VoiceId

pattern VoiceId_Mia :: VoiceId

pattern VoiceId_Miguel :: VoiceId

pattern VoiceId_Mizuki :: VoiceId

pattern VoiceId_Naja :: VoiceId

pattern VoiceId_Nicole :: VoiceId

pattern VoiceId_Olivia :: VoiceId

pattern VoiceId_Penelope :: VoiceId

pattern VoiceId_Raveena :: VoiceId

pattern VoiceId_Ricardo :: VoiceId

pattern VoiceId_Ruben :: VoiceId

pattern VoiceId_Russell :: VoiceId

pattern VoiceId_Salli :: VoiceId

pattern VoiceId_Seoyeon :: VoiceId

pattern VoiceId_Takumi :: VoiceId

pattern VoiceId_Tatyana :: VoiceId

pattern VoiceId_Vicki :: VoiceId

pattern VoiceId_Vitoria :: VoiceId

pattern VoiceId_Zeina :: VoiceId

pattern VoiceId_Zhiyu :: VoiceId

Lexicon

data Lexicon

Provides lexicon name and lexicon content in string format. For more information, see Pronunciation Lexicon Specification (PLS) Version 1.0.

See: newLexicon smart constructor.

Constructors

Lexicon' (Maybe (Sensitive Text)) (Maybe Text)

newLexicon :: Lexicon

Create a value of Lexicon with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:content:Lexicon', lexicon_content - Lexicon content in string format. The content of a lexicon must be in PLS format.

$sel:name:Lexicon', lexicon_name - Name of the lexicon.

LexiconAttributes

data LexiconAttributes

Contains metadata describing the lexicon such as the number of lexemes, language code, and so on. For more information, see Managing Lexicons.

See: newLexiconAttributes smart constructor.

Constructors

LexiconAttributes' (Maybe LanguageCode) (Maybe Int) (Maybe Int) (Maybe Text) (Maybe Text) (Maybe POSIX)

newLexiconAttributes :: LexiconAttributes

Create a value of LexiconAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:LexiconAttributes', lexiconAttributes_languageCode - Language code that the lexicon applies to. A lexicon with a language code such as "en" would be applied to all English languages (en-GB, en-US, en-AUS, en-WLS, and so on.

$sel:size:LexiconAttributes', lexiconAttributes_size - Total size of the lexicon, in characters.

$sel:lexemesCount:LexiconAttributes', lexiconAttributes_lexemesCount - Number of lexemes in the lexicon.

$sel:lexiconArn:LexiconAttributes', lexiconAttributes_lexiconArn - Amazon Resource Name (ARN) of the lexicon.

$sel:alphabet:LexiconAttributes', lexiconAttributes_alphabet - Phonetic alphabet used in the lexicon. Valid values are ipa and x-sampa.

$sel:lastModified:LexiconAttributes', lexiconAttributes_lastModified - Date lexicon was last modified (a timestamp value).

LexiconDescription

data LexiconDescription

Describes the content of the lexicon.

See: newLexiconDescription smart constructor.

Constructors

LexiconDescription' (Maybe LexiconAttributes) (Maybe Text)

newLexiconDescription :: LexiconDescription

Create a value of LexiconDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:LexiconDescription', lexiconDescription_attributes - Provides lexicon metadata.

$sel:name:LexiconDescription', lexiconDescription_name - Name of the lexicon.

SynthesisTask

data SynthesisTask

SynthesisTask object that provides information about a speech synthesis task.

See: newSynthesisTask smart constructor.

Constructors

SynthesisTask' (Maybe POSIX) (Maybe LanguageCode) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Int) (Maybe Engine) (Maybe [SpeechMarkType]) (Maybe Text) (Maybe OutputFormat) (Maybe TextType) (Maybe VoiceId) (Maybe [Text]) (Maybe TaskStatus) (Maybe Text)

newSynthesisTask :: SynthesisTask

Create a value of SynthesisTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:SynthesisTask', synthesisTask_creationTime - Timestamp for the time the synthesis task was started.

$sel:languageCode:SynthesisTask', synthesisTask_languageCode - Optional language code for a synthesis task. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly uses the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

$sel:snsTopicArn:SynthesisTask', synthesisTask_snsTopicArn - ARN for the SNS topic optionally used for providing status notification for a speech synthesis task.

$sel:taskStatusReason:SynthesisTask', synthesisTask_taskStatusReason - Reason for the current status of a specific speech synthesis task, including errors if the task has failed.

$sel:taskId:SynthesisTask', synthesisTask_taskId - The Amazon Polly generated identifier for a speech synthesis task.

$sel:requestCharacters:SynthesisTask', synthesisTask_requestCharacters - Number of billable characters synthesized.

$sel:engine:SynthesisTask', synthesisTask_engine - Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. Using a voice that is not supported for the engine selected will result in an error.

$sel:speechMarkTypes:SynthesisTask', synthesisTask_speechMarkTypes - The type of speech marks returned for the input text.

$sel:sampleRate:SynthesisTask', synthesisTask_sampleRate - The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

$sel:outputFormat:SynthesisTask', synthesisTask_outputFormat - The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

$sel:textType:SynthesisTask', synthesisTask_textType - Specifies whether the input text is plain text or SSML. The default value is plain text.

$sel:voiceId:SynthesisTask', synthesisTask_voiceId - Voice ID to use for the synthesis.

$sel:lexiconNames:SynthesisTask', synthesisTask_lexiconNames - List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice.

$sel:taskStatus:SynthesisTask', synthesisTask_taskStatus - Current status of the individual speech synthesis task.

$sel:outputUri:SynthesisTask', synthesisTask_outputUri - Pathway for the output speech file.

Voice

data Voice

Description of the voice.

See: newVoice smart constructor.

Constructors

Voice' (Maybe LanguageCode) (Maybe Text) (Maybe Gender) (Maybe Text) (Maybe VoiceId) (Maybe [LanguageCode]) (Maybe [Engine])

newVoice :: Voice

Create a value of Voice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:Voice', voice_languageCode - Language code of the voice.

$sel:languageName:Voice', voice_languageName - Human readable name of the language in English.

$sel:gender:Voice', voice_gender - Gender of the voice.

$sel:name:Voice', voice_name - Name of the voice (for example, Salli, Kendra, etc.). This provides a human readable voice name that you might display in your application.

$sel:id:Voice', voice_id - Amazon Polly assigned voice ID. This is the ID that you specify when calling the SynthesizeSpeech operation.

$sel:additionalLanguageCodes:Voice', voice_additionalLanguageCodes - Additional codes for languages available for the specified voice in addition to its default language.

For example, the default language for Aditi is Indian English (en-IN) because it was first used for that language. Since Aditi is bilingual and fluent in both Indian English and Hindi, this parameter would show the code hi-IN.

$sel:supportedEngines:Voice', voice_supportedEngines - Specifies which engines (standard or neural) that are supported by a given voice.