Amazonka.Polly.ListLexicons

Description

Returns a list of pronunciation lexicons stored in an Amazon Web Services Region. For more information, see Managing Lexicons.

This operation returns paginated results.

Creating a Request

data ListLexicons

See: newListLexicons smart constructor.

Constructors

ListLexicons'

Fields

nextToken :: Maybe Text

An opaque pagination token returned from previous ListLexicons operation. If present, indicates where to continue the list of lexicons.

newListLexicons :: ListLexicons

Create a value of ListLexicons with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLexicons', listLexicons_nextToken - An opaque pagination token returned from previous ListLexicons operation. If present, indicates where to continue the list of lexicons.

Request Lenses

listLexicons_nextToken :: Lens' ListLexicons (Maybe Text)

An opaque pagination token returned from previous ListLexicons operation. If present, indicates where to continue the list of lexicons.

Destructuring the Response

data ListLexiconsResponse

See: newListLexiconsResponse smart constructor.

Constructors

ListLexiconsResponse'

Fields

lexicons :: Maybe [LexiconDescription]

A list of lexicon names and attributes.

nextToken :: Maybe Text

The pagination token to use in the next request to continue the listing of lexicons. NextToken is returned only if the response is truncated.

httpStatus :: Int

The response's http status code.

newListLexiconsResponse

Arguments

:: Int
$sel:httpStatus:ListLexiconsResponse'

-> ListLexiconsResponse

Create a value of ListLexiconsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lexicons:ListLexiconsResponse', listLexiconsResponse_lexicons - A list of lexicon names and attributes.

$sel:nextToken:ListLexicons', listLexiconsResponse_nextToken - The pagination token to use in the next request to continue the listing of lexicons. NextToken is returned only if the response is truncated.

$sel:httpStatus:ListLexiconsResponse', listLexiconsResponse_httpStatus - The response's http status code.

Response Lenses

listLexiconsResponse_lexicons :: Lens' ListLexiconsResponse (Maybe [LexiconDescription])

A list of lexicon names and attributes.

listLexiconsResponse_nextToken :: Lens' ListLexiconsResponse (Maybe Text)

The pagination token to use in the next request to continue the listing of lexicons. NextToken is returned only if the response is truncated.

listLexiconsResponse_httpStatus :: Lens' ListLexiconsResponse Int

The response's http status code.