Amazonka.Pricing.GetAttributeValues

Description

Returns a list of attribute values. Attibutes are similar to the details in a Price List API offer file. For a list of available attributes, see Offer File Definitions in the Amazon Web Services Billing and Cost Management User Guide.

This operation returns paginated results.

Creating a Request

data GetAttributeValues

See: newGetAttributeValues smart constructor.

Constructors

GetAttributeValues'

Fields

nextToken :: Maybe Text

The pagination token that indicates the next set of results that you want to retrieve.

maxResults :: Maybe Natural

The maximum number of results to return in response.

serviceCode :: Text

The service code for the service whose attributes you want to retrieve. For example, if you want the retrieve an EC2 attribute, use AmazonEC2.

attributeName :: Text

The name of the attribute that you want to retrieve the values for, such as volumeType.

newGetAttributeValues

Arguments

:: Text
$sel:serviceCode:GetAttributeValues'

-> Text
$sel:attributeName:GetAttributeValues'

-> GetAttributeValues

Create a value of GetAttributeValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetAttributeValues', getAttributeValues_nextToken - The pagination token that indicates the next set of results that you want to retrieve.

$sel:maxResults:GetAttributeValues', getAttributeValues_maxResults - The maximum number of results to return in response.

$sel:serviceCode:GetAttributeValues', getAttributeValues_serviceCode - The service code for the service whose attributes you want to retrieve. For example, if you want the retrieve an EC2 attribute, use AmazonEC2.

$sel:attributeName:GetAttributeValues', getAttributeValues_attributeName - The name of the attribute that you want to retrieve the values for, such as volumeType.

Request Lenses

getAttributeValues_nextToken :: Lens' GetAttributeValues (Maybe Text)

The pagination token that indicates the next set of results that you want to retrieve.

getAttributeValues_maxResults :: Lens' GetAttributeValues (Maybe Natural)

The maximum number of results to return in response.

getAttributeValues_serviceCode :: Lens' GetAttributeValues Text

The service code for the service whose attributes you want to retrieve. For example, if you want the retrieve an EC2 attribute, use AmazonEC2.

getAttributeValues_attributeName :: Lens' GetAttributeValues Text

The name of the attribute that you want to retrieve the values for, such as volumeType.

Destructuring the Response

data GetAttributeValuesResponse

See: newGetAttributeValuesResponse smart constructor.

Constructors

GetAttributeValuesResponse'

Fields

attributeValues :: Maybe [AttributeValue]

The list of values for an attribute. For example, Throughput Optimized HDD and Provisioned IOPS are two available values for the AmazonEC2 volumeType.

nextToken :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

httpStatus :: Int

The response's http status code.

newGetAttributeValuesResponse

Arguments

:: Int
$sel:httpStatus:GetAttributeValuesResponse'

-> GetAttributeValuesResponse

Create a value of GetAttributeValuesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeValues:GetAttributeValuesResponse', getAttributeValuesResponse_attributeValues - The list of values for an attribute. For example, Throughput Optimized HDD and Provisioned IOPS are two available values for the AmazonEC2 volumeType.

$sel:nextToken:GetAttributeValues', getAttributeValuesResponse_nextToken - The pagination token that indicates the next set of results to retrieve.

$sel:httpStatus:GetAttributeValuesResponse', getAttributeValuesResponse_httpStatus - The response's http status code.

Response Lenses

getAttributeValuesResponse_attributeValues :: Lens' GetAttributeValuesResponse (Maybe [AttributeValue])

The list of values for an attribute. For example, Throughput Optimized HDD and Provisioned IOPS are two available values for the AmazonEC2 volumeType.

getAttributeValuesResponse_nextToken :: Lens' GetAttributeValuesResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

getAttributeValuesResponse_httpStatus :: Lens' GetAttributeValuesResponse Int

The response's http status code.