Amazonka.Pricing.Lens

Description

Operations

GetAttributeValues

getAttributeValues_nextToken :: Lens' GetAttributeValues (Maybe Text)

The pagination token that indicates the next set of results that you want to retrieve.

getAttributeValues_maxResults :: Lens' GetAttributeValues (Maybe Natural)

The maximum number of results to return in response.

getAttributeValues_serviceCode :: Lens' GetAttributeValues Text

The service code for the service whose attributes you want to retrieve. For example, if you want the retrieve an EC2 attribute, use AmazonEC2.

getAttributeValues_attributeName :: Lens' GetAttributeValues Text

The name of the attribute that you want to retrieve the values for, such as volumeType.

getAttributeValuesResponse_attributeValues :: Lens' GetAttributeValuesResponse (Maybe [AttributeValue])

The list of values for an attribute. For example, Throughput Optimized HDD and Provisioned IOPS are two available values for the AmazonEC2 volumeType.

getAttributeValuesResponse_nextToken :: Lens' GetAttributeValuesResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

getAttributeValuesResponse_httpStatus :: Lens' GetAttributeValuesResponse Int

The response's http status code.

DescribeServices

describeServices_formatVersion :: Lens' DescribeServices (Maybe Text)

The format version that you want the response to be in.

Valid values are: aws_v1

describeServices_nextToken :: Lens' DescribeServices (Maybe Text)

The pagination token that indicates the next set of results that you want to retrieve.

describeServices_serviceCode :: Lens' DescribeServices (Maybe Text)

The code for the service whose information you want to retrieve, such as AmazonEC2. You can use the ServiceCode to filter the results in a GetProducts call. To retrieve a list of all services, leave this blank.

describeServices_maxResults :: Lens' DescribeServices (Maybe Natural)

The maximum number of results that you want returned in the response.

describeServicesResponse_formatVersion :: Lens' DescribeServicesResponse (Maybe Text)

The format version of the response. For example, aws_v1.

describeServicesResponse_nextToken :: Lens' DescribeServicesResponse (Maybe Text)

The pagination token for the next set of retreivable results.

describeServicesResponse_services :: Lens' DescribeServicesResponse (Maybe [PricingService])

The service metadata for the service or services in the response.

describeServicesResponse_httpStatus :: Lens' DescribeServicesResponse Int

The response's http status code.

GetProducts

getProducts_filters :: Lens' GetProducts (Maybe [Filter])

The list of filters that limit the returned products. only products that match all filters are returned.

getProducts_formatVersion :: Lens' GetProducts (Maybe Text)

The format version that you want the response to be in.

Valid values are: aws_v1

getProducts_nextToken :: Lens' GetProducts (Maybe Text)

The pagination token that indicates the next set of results that you want to retrieve.

getProducts_serviceCode :: Lens' GetProducts (Maybe Text)

The code for the service whose products you want to retrieve.

getProducts_maxResults :: Lens' GetProducts (Maybe Natural)

The maximum number of results to return in the response.

getProductsResponse_formatVersion :: Lens' GetProductsResponse (Maybe Text)

The format version of the response. For example, aws_v1.

getProductsResponse_nextToken :: Lens' GetProductsResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

getProductsResponse_priceList :: Lens' GetProductsResponse (Maybe [Text])

The list of products that match your filters. The list contains both the product metadata and the price information.

getProductsResponse_httpStatus :: Lens' GetProductsResponse Int

The response's http status code.

Types

AttributeValue

attributeValue_value :: Lens' AttributeValue (Maybe Text)

The specific value of an attributeName.

Filter

filter_type :: Lens' Filter FilterType

The type of filter that you want to use.

Valid values are: TERM_MATCH. TERM_MATCH returns only products that match both the given filter field and the given value.

filter_field :: Lens' Filter Text

The product metadata field that you want to filter on. You can filter by just the service code to see all products for a specific service, filter by just the attribute name to see a specific attribute for multiple services, or use both a service code and an attribute name to retrieve only products that match both fields.

Valid values include: ServiceCode, and all attribute names

For example, you can filter by the AmazonEC2 service code and the volumeType attribute name to get the prices for only Amazon EC2 volumes.

filter_value :: Lens' Filter Text

The service code or attribute value that you want to filter by. If you are filtering by service code this is the actual service code, such as AmazonEC2. If you are filtering by attribute name, this is the attribute value that you want the returned products to match, such as a Provisioned IOPS volume.

PricingService

pricingService_attributeNames :: Lens' PricingService (Maybe [Text])

The attributes that are available for this service.

pricingService_serviceCode :: Lens' PricingService (Maybe Text)

The code for the Amazon Web Services service.