Amazonka.Proton.CreateEnvironment

Description

Deploy a new environment. An AWS Proton environment is created from an environment template that defines infrastructure and resources that can be shared across services. For more information, see the Environments in the AWS Proton Administrator Guide.

Creating a Request

data CreateEnvironment

See: newCreateEnvironment smart constructor.

Constructors

CreateEnvironment'

Fields

protonServiceRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS Proton service role that allows AWS Proton to make calls to other services on your behalf. You must include either the environmentAccountConnectionId or protonServiceRoleArn parameter and value.

environmentAccountConnectionId :: Maybe Text

The ID of the environment account connection that you provide if you're provisioning your environment infrastructure resources to an environment account. You must include either the environmentAccountConnectionId or protonServiceRoleArn parameter and value. For more information, see Environment account connections in the AWS Proton Administrator guide.

templateMinorVersion :: Maybe Text

The ID of the minor version of the environment template.

description :: Maybe (Sensitive Text)

A description of the environment that's being created and deployed.

tags :: Maybe [Tag]

Create tags for your environment. For more information, see /AWS Proton resources and tagging/ in the AWS Proton Administrator Guide or AWS Proton User Guide.

name :: Text

The name of the environment.

spec :: Sensitive Text

A link to a YAML formatted spec file that provides inputs as defined in the environment template bundle schema file. For more information, see Environments in the AWS Proton Administrator Guide.

templateMajorVersion :: Text

The ID of the major version of the environment template.

templateName :: Text

The name of the environment template. For more information, see Environment Templates in the AWS Proton Administrator Guide.

newCreateEnvironment

Arguments

:: Text
$sel:name:CreateEnvironment'

-> Text
$sel:spec:CreateEnvironment'

-> Text
$sel:templateMajorVersion:CreateEnvironment'

-> Text
$sel:templateName:CreateEnvironment'

-> CreateEnvironment

Create a value of CreateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protonServiceRoleArn:CreateEnvironment', createEnvironment_protonServiceRoleArn - The Amazon Resource Name (ARN) of the AWS Proton service role that allows AWS Proton to make calls to other services on your behalf. You must include either the environmentAccountConnectionId or protonServiceRoleArn parameter and value.

$sel:environmentAccountConnectionId:CreateEnvironment', createEnvironment_environmentAccountConnectionId - The ID of the environment account connection that you provide if you're provisioning your environment infrastructure resources to an environment account. You must include either the environmentAccountConnectionId or protonServiceRoleArn parameter and value. For more information, see Environment account connections in the AWS Proton Administrator guide.

$sel:templateMinorVersion:CreateEnvironment', createEnvironment_templateMinorVersion - The ID of the minor version of the environment template.

$sel:description:CreateEnvironment', createEnvironment_description - A description of the environment that's being created and deployed.

$sel:tags:CreateEnvironment', createEnvironment_tags - Create tags for your environment. For more information, see /AWS Proton resources and tagging/ in the AWS Proton Administrator Guide or AWS Proton User Guide.

$sel:name:CreateEnvironment', createEnvironment_name - The name of the environment.

$sel:spec:CreateEnvironment', createEnvironment_spec - A link to a YAML formatted spec file that provides inputs as defined in the environment template bundle schema file. For more information, see Environments in the AWS Proton Administrator Guide.

$sel:templateMajorVersion:CreateEnvironment', createEnvironment_templateMajorVersion - The ID of the major version of the environment template.

$sel:templateName:CreateEnvironment', createEnvironment_templateName - The name of the environment template. For more information, see Environment Templates in the AWS Proton Administrator Guide.

Request Lenses

createEnvironment_protonServiceRoleArn :: Lens' CreateEnvironment (Maybe Text)

The Amazon Resource Name (ARN) of the AWS Proton service role that allows AWS Proton to make calls to other services on your behalf. You must include either the environmentAccountConnectionId or protonServiceRoleArn parameter and value.

createEnvironment_environmentAccountConnectionId :: Lens' CreateEnvironment (Maybe Text)

The ID of the environment account connection that you provide if you're provisioning your environment infrastructure resources to an environment account. You must include either the environmentAccountConnectionId or protonServiceRoleArn parameter and value. For more information, see Environment account connections in the AWS Proton Administrator guide.

createEnvironment_templateMinorVersion :: Lens' CreateEnvironment (Maybe Text)

The ID of the minor version of the environment template.

createEnvironment_description :: Lens' CreateEnvironment (Maybe Text)

A description of the environment that's being created and deployed.

createEnvironment_tags :: Lens' CreateEnvironment (Maybe [Tag])

Create tags for your environment. For more information, see /AWS Proton resources and tagging/ in the AWS Proton Administrator Guide or AWS Proton User Guide.

createEnvironment_name :: Lens' CreateEnvironment Text

The name of the environment.

createEnvironment_spec :: Lens' CreateEnvironment Text

A link to a YAML formatted spec file that provides inputs as defined in the environment template bundle schema file. For more information, see Environments in the AWS Proton Administrator Guide.

createEnvironment_templateMajorVersion :: Lens' CreateEnvironment Text

The ID of the major version of the environment template.

createEnvironment_templateName :: Lens' CreateEnvironment Text

The name of the environment template. For more information, see Environment Templates in the AWS Proton Administrator Guide.

Destructuring the Response

data CreateEnvironmentResponse

See: newCreateEnvironmentResponse smart constructor.

Constructors

CreateEnvironmentResponse'

Fields

httpStatus :: Int

The response's http status code.

environment :: Environment

The environment detail data that's returned by AWS Proton.

newCreateEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:CreateEnvironmentResponse'

-> Environment
$sel:environment:CreateEnvironmentResponse'

-> CreateEnvironmentResponse

Create a value of CreateEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateEnvironmentResponse', createEnvironmentResponse_httpStatus - The response's http status code.

$sel:environment:CreateEnvironmentResponse', createEnvironmentResponse_environment - The environment detail data that's returned by AWS Proton.

Response Lenses

createEnvironmentResponse_httpStatus :: Lens' CreateEnvironmentResponse Int

The response's http status code.

createEnvironmentResponse_environment :: Lens' CreateEnvironmentResponse Environment

The environment detail data that's returned by AWS Proton.