Amazonka.Proton.CreateEnvironmentAccountConnection

Description

Create an environment account connection in an environment account so that environment infrastructure resources can be provisioned in the environment account from a management account.

An environment account connection is a secure bi-directional connection between a management account and an environment account that maintains authorization and permissions. For more information, see Environment account connections in the AWS Proton Administrator guide.

Creating a Request

data CreateEnvironmentAccountConnection

See: newCreateEnvironmentAccountConnection smart constructor.

Constructors

CreateEnvironmentAccountConnection'

Fields

clientToken :: Maybe Text

When included, if two identicial requests are made with the same client token, AWS Proton returns the environment account connection that the first request created.

environmentName :: Text

The name of the AWS Proton environment that's created in the associated management account.

managementAccountId :: Text

The ID of the management account that accepts or rejects the environment account connection. You create an manage the AWS Proton environment in this account. If the management account accepts the environment account connection, AWS Proton can use the associated IAM role to provision environment infrastructure resources in the associated environment account.

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM service role that's created in the environment account. AWS Proton uses this role to provision infrastructure resources in the associated environment account.

newCreateEnvironmentAccountConnection

Arguments

:: Text
$sel:environmentName:CreateEnvironmentAccountConnection'

-> Text
$sel:managementAccountId:CreateEnvironmentAccountConnection'

-> Text
$sel:roleArn:CreateEnvironmentAccountConnection'

-> CreateEnvironmentAccountConnection

Create a value of CreateEnvironmentAccountConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateEnvironmentAccountConnection', createEnvironmentAccountConnection_clientToken - When included, if two identicial requests are made with the same client token, AWS Proton returns the environment account connection that the first request created.

$sel:environmentName:CreateEnvironmentAccountConnection', createEnvironmentAccountConnection_environmentName - The name of the AWS Proton environment that's created in the associated management account.

$sel:managementAccountId:CreateEnvironmentAccountConnection', createEnvironmentAccountConnection_managementAccountId - The ID of the management account that accepts or rejects the environment account connection. You create an manage the AWS Proton environment in this account. If the management account accepts the environment account connection, AWS Proton can use the associated IAM role to provision environment infrastructure resources in the associated environment account.

$sel:roleArn:CreateEnvironmentAccountConnection', createEnvironmentAccountConnection_roleArn - The Amazon Resource Name (ARN) of the IAM service role that's created in the environment account. AWS Proton uses this role to provision infrastructure resources in the associated environment account.

Request Lenses

createEnvironmentAccountConnection_clientToken :: Lens' CreateEnvironmentAccountConnection (Maybe Text)

When included, if two identicial requests are made with the same client token, AWS Proton returns the environment account connection that the first request created.

createEnvironmentAccountConnection_environmentName :: Lens' CreateEnvironmentAccountConnection Text

The name of the AWS Proton environment that's created in the associated management account.

createEnvironmentAccountConnection_managementAccountId :: Lens' CreateEnvironmentAccountConnection Text

The ID of the management account that accepts or rejects the environment account connection. You create an manage the AWS Proton environment in this account. If the management account accepts the environment account connection, AWS Proton can use the associated IAM role to provision environment infrastructure resources in the associated environment account.

createEnvironmentAccountConnection_roleArn :: Lens' CreateEnvironmentAccountConnection Text

The Amazon Resource Name (ARN) of the IAM service role that's created in the environment account. AWS Proton uses this role to provision infrastructure resources in the associated environment account.

Destructuring the Response

data CreateEnvironmentAccountConnectionResponse

See: newCreateEnvironmentAccountConnectionResponse smart constructor.

Constructors

CreateEnvironmentAccountConnectionResponse'

Fields

httpStatus :: Int

The response's http status code.

environmentAccountConnection :: EnvironmentAccountConnection

The environment account connection detail data that's returned by AWS Proton.

newCreateEnvironmentAccountConnectionResponse

Arguments

:: Int
$sel:httpStatus:CreateEnvironmentAccountConnectionResponse'

-> EnvironmentAccountConnection
$sel:environmentAccountConnection:CreateEnvironmentAccountConnectionResponse'

-> CreateEnvironmentAccountConnectionResponse

Create a value of CreateEnvironmentAccountConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateEnvironmentAccountConnectionResponse', createEnvironmentAccountConnectionResponse_httpStatus - The response's http status code.

$sel:environmentAccountConnection:CreateEnvironmentAccountConnectionResponse', createEnvironmentAccountConnectionResponse_environmentAccountConnection - The environment account connection detail data that's returned by AWS Proton.

Response Lenses

createEnvironmentAccountConnectionResponse_httpStatus :: Lens' CreateEnvironmentAccountConnectionResponse Int

The response's http status code.

createEnvironmentAccountConnectionResponse_environmentAccountConnection :: Lens' CreateEnvironmentAccountConnectionResponse EnvironmentAccountConnection

The environment account connection detail data that's returned by AWS Proton.