Amazonka.Proton.CreateService

Description

Create an AWS Proton service. An AWS Proton service is an instantiation of a service template and often includes several service instances and pipeline. For more information, see Services in the AWS Proton Administrator Guide and Services in the AWS Proton User Guide.

Creating a Request

data CreateService

See: newCreateService smart constructor.

Constructors

CreateService'

Fields

branchName :: Maybe Text

The name of the code repository branch that holds the code that's deployed in AWS Proton. Don't include this parameter if your service template doesn't include a service pipeline.

repositoryId :: Maybe Text

The ID of the code repository. Don't include this parameter if your service template doesn't include a service pipeline.

templateMinorVersion :: Maybe Text

The ID of the minor version of the service template that was used to create the service.

description :: Maybe (Sensitive Text)

A description of the AWS Proton service.

repositoryConnectionArn :: Maybe Text

The Amazon Resource Name (ARN) of the repository connection. For more information, see Set up repository connection in the AWS Proton Administrator Guide and Setting up with AWS Proton in the AWS Proton User Guide. Don't include this parameter if your service template doesn't include a service pipeline.

tags :: Maybe [Tag]

Create tags for your service. For more information, see /AWS Proton resources and tagging/ in the AWS Proton Administrator Guide or AWS Proton User Guide.

name :: Text

The service name.

spec :: Sensitive Text

A link to a spec file that provides inputs as defined in the service template bundle schema file. The spec file is in YAML format. Don’t include pipeline inputs in the spec if your service template doesn’t include a service pipeline. For more information, see Create a service in the AWS Proton Administrator Guide and Create a service in the AWS Proton User Guide.

templateMajorVersion :: Text

The ID of the major version of the service template that was used to create the service.

templateName :: Text

The name of the service template that's used to create the service.

newCreateService

Arguments

:: Text
$sel:name:CreateService'

-> Text
$sel:spec:CreateService'

-> Text
$sel:templateMajorVersion:CreateService'

-> Text
$sel:templateName:CreateService'

-> CreateService

Create a value of CreateService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:branchName:CreateService', createService_branchName - The name of the code repository branch that holds the code that's deployed in AWS Proton. Don't include this parameter if your service template doesn't include a service pipeline.

$sel:repositoryId:CreateService', createService_repositoryId - The ID of the code repository. Don't include this parameter if your service template doesn't include a service pipeline.

$sel:templateMinorVersion:CreateService', createService_templateMinorVersion - The ID of the minor version of the service template that was used to create the service.

$sel:description:CreateService', createService_description - A description of the AWS Proton service.

$sel:repositoryConnectionArn:CreateService', createService_repositoryConnectionArn - The Amazon Resource Name (ARN) of the repository connection. For more information, see Set up repository connection in the AWS Proton Administrator Guide and Setting up with AWS Proton in the AWS Proton User Guide. Don't include this parameter if your service template doesn't include a service pipeline.

$sel:tags:CreateService', createService_tags - Create tags for your service. For more information, see /AWS Proton resources and tagging/ in the AWS Proton Administrator Guide or AWS Proton User Guide.

$sel:name:CreateService', createService_name - The service name.

$sel:spec:CreateService', createService_spec - A link to a spec file that provides inputs as defined in the service template bundle schema file. The spec file is in YAML format. Don’t include pipeline inputs in the spec if your service template doesn’t include a service pipeline. For more information, see Create a service in the AWS Proton Administrator Guide and Create a service in the AWS Proton User Guide.

$sel:templateMajorVersion:CreateService', createService_templateMajorVersion - The ID of the major version of the service template that was used to create the service.

$sel:templateName:CreateService', createService_templateName - The name of the service template that's used to create the service.

Request Lenses

createService_branchName :: Lens' CreateService (Maybe Text)

The name of the code repository branch that holds the code that's deployed in AWS Proton. Don't include this parameter if your service template doesn't include a service pipeline.

createService_repositoryId :: Lens' CreateService (Maybe Text)

The ID of the code repository. Don't include this parameter if your service template doesn't include a service pipeline.

createService_templateMinorVersion :: Lens' CreateService (Maybe Text)

The ID of the minor version of the service template that was used to create the service.

createService_description :: Lens' CreateService (Maybe Text)

A description of the AWS Proton service.

createService_repositoryConnectionArn :: Lens' CreateService (Maybe Text)

The Amazon Resource Name (ARN) of the repository connection. For more information, see Set up repository connection in the AWS Proton Administrator Guide and Setting up with AWS Proton in the AWS Proton User Guide. Don't include this parameter if your service template doesn't include a service pipeline.

createService_tags :: Lens' CreateService (Maybe [Tag])

Create tags for your service. For more information, see /AWS Proton resources and tagging/ in the AWS Proton Administrator Guide or AWS Proton User Guide.

createService_name :: Lens' CreateService Text

The service name.

createService_spec :: Lens' CreateService Text

A link to a spec file that provides inputs as defined in the service template bundle schema file. The spec file is in YAML format. Don’t include pipeline inputs in the spec if your service template doesn’t include a service pipeline. For more information, see Create a service in the AWS Proton Administrator Guide and Create a service in the AWS Proton User Guide.

createService_templateMajorVersion :: Lens' CreateService Text

The ID of the major version of the service template that was used to create the service.

createService_templateName :: Lens' CreateService Text

The name of the service template that's used to create the service.

Destructuring the Response

data CreateServiceResponse

See: newCreateServiceResponse smart constructor.

Constructors

CreateServiceResponse'

Fields

httpStatus :: Int

The response's http status code.

service :: Service

The service detail data that's returned by AWS Proton.

newCreateServiceResponse

Arguments

:: Int
$sel:httpStatus:CreateServiceResponse'

-> Service
$sel:service:CreateServiceResponse'

-> CreateServiceResponse

Create a value of CreateServiceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateServiceResponse', createServiceResponse_httpStatus - The response's http status code.

$sel:service:CreateServiceResponse', createServiceResponse_service - The service detail data that's returned by AWS Proton.

Response Lenses

createServiceResponse_httpStatus :: Lens' CreateServiceResponse Int

The response's http status code.

createServiceResponse_service :: Lens' CreateServiceResponse Service

The service detail data that's returned by AWS Proton.