Amazonka.Proton.CreateServiceTemplateVersion

Description

Create a new major or minor version of a service template. A major version of a service template is a version that isn't backwards compatible. A minor version of a service template is a version that's backwards compatible within its major version.

Creating a Request

data CreateServiceTemplateVersion

See: newCreateServiceTemplateVersion smart constructor.

Constructors

CreateServiceTemplateVersion'

Fields

clientToken :: Maybe Text

When included, if two identicial requests are made with the same client token, AWS Proton returns the service template version that the first request created.

majorVersion :: Maybe Text

To create a new minor version of the service template, include a majorVersion.

To create a new major and minor version of the service template, exclude majorVersion.

description :: Maybe (Sensitive Text)

A description of the new version of a service template.

tags :: Maybe [Tag]

Create tags for a new version of a service template.

compatibleEnvironmentTemplates :: NonEmpty CompatibleEnvironmentTemplateInput

An array of compatible environment template objects for the new version of a service template.

source :: TemplateVersionSourceInput

An object that includes the template bundle S3 bucket path and name for the new version of a service template.

templateName :: Text

The name of the service template.

newCreateServiceTemplateVersion

Arguments

:: NonEmpty CompatibleEnvironmentTemplateInput
$sel:compatibleEnvironmentTemplates:CreateServiceTemplateVersion'

-> TemplateVersionSourceInput
$sel:source:CreateServiceTemplateVersion'

-> Text
$sel:templateName:CreateServiceTemplateVersion'

-> CreateServiceTemplateVersion

Create a value of CreateServiceTemplateVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateServiceTemplateVersion', createServiceTemplateVersion_clientToken - When included, if two identicial requests are made with the same client token, AWS Proton returns the service template version that the first request created.

$sel:majorVersion:CreateServiceTemplateVersion', createServiceTemplateVersion_majorVersion - To create a new minor version of the service template, include a majorVersion.

To create a new major and minor version of the service template, exclude majorVersion.

$sel:description:CreateServiceTemplateVersion', createServiceTemplateVersion_description - A description of the new version of a service template.

$sel:tags:CreateServiceTemplateVersion', createServiceTemplateVersion_tags - Create tags for a new version of a service template.

$sel:compatibleEnvironmentTemplates:CreateServiceTemplateVersion', createServiceTemplateVersion_compatibleEnvironmentTemplates - An array of compatible environment template objects for the new version of a service template.

$sel:source:CreateServiceTemplateVersion', createServiceTemplateVersion_source - An object that includes the template bundle S3 bucket path and name for the new version of a service template.

$sel:templateName:CreateServiceTemplateVersion', createServiceTemplateVersion_templateName - The name of the service template.

Request Lenses

createServiceTemplateVersion_clientToken :: Lens' CreateServiceTemplateVersion (Maybe Text)

When included, if two identicial requests are made with the same client token, AWS Proton returns the service template version that the first request created.

createServiceTemplateVersion_majorVersion :: Lens' CreateServiceTemplateVersion (Maybe Text)

To create a new minor version of the service template, include a majorVersion.

To create a new major and minor version of the service template, exclude majorVersion.

createServiceTemplateVersion_description :: Lens' CreateServiceTemplateVersion (Maybe Text)

A description of the new version of a service template.

createServiceTemplateVersion_tags :: Lens' CreateServiceTemplateVersion (Maybe [Tag])

Create tags for a new version of a service template.

createServiceTemplateVersion_compatibleEnvironmentTemplates :: Lens' CreateServiceTemplateVersion (NonEmpty CompatibleEnvironmentTemplateInput)

An array of compatible environment template objects for the new version of a service template.

createServiceTemplateVersion_source :: Lens' CreateServiceTemplateVersion TemplateVersionSourceInput

An object that includes the template bundle S3 bucket path and name for the new version of a service template.

createServiceTemplateVersion_templateName :: Lens' CreateServiceTemplateVersion Text

The name of the service template.

Destructuring the Response

data CreateServiceTemplateVersionResponse

See: newCreateServiceTemplateVersionResponse smart constructor.

Constructors

CreateServiceTemplateVersionResponse'

Fields

httpStatus :: Int

The response's http status code.

serviceTemplateVersion :: ServiceTemplateVersion

The service template version summary of detail data that's returned by AWS Proton.

newCreateServiceTemplateVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateServiceTemplateVersionResponse'

-> ServiceTemplateVersion
$sel:serviceTemplateVersion:CreateServiceTemplateVersionResponse'

-> CreateServiceTemplateVersionResponse

Create a value of CreateServiceTemplateVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateServiceTemplateVersionResponse', createServiceTemplateVersionResponse_httpStatus - The response's http status code.

$sel:serviceTemplateVersion:CreateServiceTemplateVersionResponse', createServiceTemplateVersionResponse_serviceTemplateVersion - The service template version summary of detail data that's returned by AWS Proton.

Response Lenses

createServiceTemplateVersionResponse_httpStatus :: Lens' CreateServiceTemplateVersionResponse Int

The response's http status code.

createServiceTemplateVersionResponse_serviceTemplateVersion :: Lens' CreateServiceTemplateVersionResponse ServiceTemplateVersion

The service template version summary of detail data that's returned by AWS Proton.