Amazonka.Proton.DeleteServiceTemplateVersion

Description

If no other minor versions of a service template exist, delete a major version of the service template if it's not the Recommended version. Delete the Recommended version of the service template if no other major versions or minor versions of the service template exist. A major version of a service template is a version that isn't backwards compatible.

Delete a minor version of a service template if it's not the Recommended version. Delete a Recommended minor version of the service template if no other minor versions of the service template exist. A minor version of a service template is a version that's backwards compatible.

Creating a Request

data DeleteServiceTemplateVersion

See: newDeleteServiceTemplateVersion smart constructor.

Constructors

DeleteServiceTemplateVersion'

Fields

majorVersion :: Text

The service template major version to delete.

minorVersion :: Text

The service template minor version to delete.

templateName :: Text

The name of the service template.

newDeleteServiceTemplateVersion

Arguments

:: Text
$sel:majorVersion:DeleteServiceTemplateVersion'

-> Text
$sel:minorVersion:DeleteServiceTemplateVersion'

-> Text
$sel:templateName:DeleteServiceTemplateVersion'

-> DeleteServiceTemplateVersion

Create a value of DeleteServiceTemplateVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:majorVersion:DeleteServiceTemplateVersion', deleteServiceTemplateVersion_majorVersion - The service template major version to delete.

$sel:minorVersion:DeleteServiceTemplateVersion', deleteServiceTemplateVersion_minorVersion - The service template minor version to delete.

$sel:templateName:DeleteServiceTemplateVersion', deleteServiceTemplateVersion_templateName - The name of the service template.

Request Lenses

deleteServiceTemplateVersion_majorVersion :: Lens' DeleteServiceTemplateVersion Text

The service template major version to delete.

deleteServiceTemplateVersion_minorVersion :: Lens' DeleteServiceTemplateVersion Text

The service template minor version to delete.

deleteServiceTemplateVersion_templateName :: Lens' DeleteServiceTemplateVersion Text

The name of the service template.

Destructuring the Response

data DeleteServiceTemplateVersionResponse

See: newDeleteServiceTemplateVersionResponse smart constructor.

Constructors

DeleteServiceTemplateVersionResponse'

Fields

serviceTemplateVersion :: Maybe ServiceTemplateVersion

The service template version detail data that's returned by AWS Proton.

httpStatus :: Int

The response's http status code.

newDeleteServiceTemplateVersionResponse

Arguments

:: Int
$sel:httpStatus:DeleteServiceTemplateVersionResponse'

-> DeleteServiceTemplateVersionResponse

Create a value of DeleteServiceTemplateVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceTemplateVersion:DeleteServiceTemplateVersionResponse', deleteServiceTemplateVersionResponse_serviceTemplateVersion - The service template version detail data that's returned by AWS Proton.

$sel:httpStatus:DeleteServiceTemplateVersionResponse', deleteServiceTemplateVersionResponse_httpStatus - The response's http status code.

Response Lenses

deleteServiceTemplateVersionResponse_serviceTemplateVersion :: Lens' DeleteServiceTemplateVersionResponse (Maybe ServiceTemplateVersion)

The service template version detail data that's returned by AWS Proton.

deleteServiceTemplateVersionResponse_httpStatus :: Lens' DeleteServiceTemplateVersionResponse Int

The response's http status code.