Amazonka.Proton.GetEnvironmentTemplate

Description

Get detail data for an environment template.

Creating a Request

data GetEnvironmentTemplate

See: newGetEnvironmentTemplate smart constructor.

Constructors

GetEnvironmentTemplate'

Fields

name :: Text

The name of the environment template that you want to get the detail data for.

newGetEnvironmentTemplate

Arguments

:: Text
$sel:name:GetEnvironmentTemplate'

-> GetEnvironmentTemplate

Create a value of GetEnvironmentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetEnvironmentTemplate', getEnvironmentTemplate_name - The name of the environment template that you want to get the detail data for.

Request Lenses

getEnvironmentTemplate_name :: Lens' GetEnvironmentTemplate Text

The name of the environment template that you want to get the detail data for.

Destructuring the Response

data GetEnvironmentTemplateResponse

See: newGetEnvironmentTemplateResponse smart constructor.

Constructors

GetEnvironmentTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

environmentTemplate :: EnvironmentTemplate

The environment template detail data that's returned by AWS Proton.

newGetEnvironmentTemplateResponse

Arguments

:: Int
$sel:httpStatus:GetEnvironmentTemplateResponse'

-> EnvironmentTemplate
$sel:environmentTemplate:GetEnvironmentTemplateResponse'

-> GetEnvironmentTemplateResponse

Create a value of GetEnvironmentTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetEnvironmentTemplateResponse', getEnvironmentTemplateResponse_httpStatus - The response's http status code.

$sel:environmentTemplate:GetEnvironmentTemplateResponse', getEnvironmentTemplateResponse_environmentTemplate - The environment template detail data that's returned by AWS Proton.

Response Lenses

getEnvironmentTemplateResponse_httpStatus :: Lens' GetEnvironmentTemplateResponse Int

The response's http status code.

getEnvironmentTemplateResponse_environmentTemplate :: Lens' GetEnvironmentTemplateResponse EnvironmentTemplate

The environment template detail data that's returned by AWS Proton.