Amazonka.Proton.Types.Environment

Description

Documentation

data Environment

The environment detail data. An AWS Proton environment is a set resources shared across an AWS Proton service.

See: newEnvironment smart constructor.

Constructors

Environment'

Fields

deploymentStatusMessage :: Maybe (Sensitive Text)

An environment deployment status message.

environmentAccountId :: Maybe Text

The ID of the environment account that the environment infrastructure resources are provisioned in.

provisioning :: Maybe Provisioning

When included, indicates that the environment template is for customer provisioned and managed infrastructure.

protonServiceRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS Proton service role that allows AWS Proton to make calls to other services on your behalf.

environmentAccountConnectionId :: Maybe Text

The ID of the environment account connection that's used to provision infrastructure resources in an environment account.

spec :: Maybe (Sensitive Text)

The environment spec.

description :: Maybe (Sensitive Text)

The description of the environment.

arn :: Text

The Amazon Resource Name (ARN) of the environment.

createdAt :: POSIX

The time when the environment was created.

deploymentStatus :: DeploymentStatus

The environment deployment status.

lastDeploymentAttemptedAt :: POSIX

The time when a deployment of the environment was last attempted.

lastDeploymentSucceededAt :: POSIX

The time when the environment was last deployed successfully.

name :: Text

The name of the environment.

templateMajorVersion :: Text

The ID of the major version of the environment template.

templateMinorVersion :: Text

The ID of the minor version of the environment template.

templateName :: Text

The Amazon Resource Name (ARN) of the environment template.

newEnvironment

Arguments

:: Text
$sel:arn:Environment'

-> UTCTime
$sel:createdAt:Environment'

-> DeploymentStatus
$sel:deploymentStatus:Environment'

-> UTCTime
$sel:lastDeploymentAttemptedAt:Environment'

-> UTCTime
$sel:lastDeploymentSucceededAt:Environment'

-> Text
$sel:name:Environment'

-> Text
$sel:templateMajorVersion:Environment'

-> Text
$sel:templateMinorVersion:Environment'

-> Text
$sel:templateName:Environment'

-> Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentStatusMessage:Environment', environment_deploymentStatusMessage - An environment deployment status message.

$sel:environmentAccountId:Environment', environment_environmentAccountId - The ID of the environment account that the environment infrastructure resources are provisioned in.

$sel:provisioning:Environment', environment_provisioning - When included, indicates that the environment template is for customer provisioned and managed infrastructure.

$sel:protonServiceRoleArn:Environment', environment_protonServiceRoleArn - The Amazon Resource Name (ARN) of the AWS Proton service role that allows AWS Proton to make calls to other services on your behalf.

$sel:environmentAccountConnectionId:Environment', environment_environmentAccountConnectionId - The ID of the environment account connection that's used to provision infrastructure resources in an environment account.

$sel:spec:Environment', environment_spec - The environment spec.

$sel:description:Environment', environment_description - The description of the environment.

$sel:arn:Environment', environment_arn - The Amazon Resource Name (ARN) of the environment.

$sel:createdAt:Environment', environment_createdAt - The time when the environment was created.

$sel:deploymentStatus:Environment', environment_deploymentStatus - The environment deployment status.

$sel:lastDeploymentAttemptedAt:Environment', environment_lastDeploymentAttemptedAt - The time when a deployment of the environment was last attempted.

$sel:lastDeploymentSucceededAt:Environment', environment_lastDeploymentSucceededAt - The time when the environment was last deployed successfully.

$sel:name:Environment', environment_name - The name of the environment.

$sel:templateMajorVersion:Environment', environment_templateMajorVersion - The ID of the major version of the environment template.

$sel:templateMinorVersion:Environment', environment_templateMinorVersion - The ID of the minor version of the environment template.

$sel:templateName:Environment', environment_templateName - The Amazon Resource Name (ARN) of the environment template.

environment_deploymentStatusMessage :: Lens' Environment (Maybe Text)

An environment deployment status message.

environment_environmentAccountId :: Lens' Environment (Maybe Text)

The ID of the environment account that the environment infrastructure resources are provisioned in.

environment_provisioning :: Lens' Environment (Maybe Provisioning)

When included, indicates that the environment template is for customer provisioned and managed infrastructure.

environment_protonServiceRoleArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of the AWS Proton service role that allows AWS Proton to make calls to other services on your behalf.

environment_environmentAccountConnectionId :: Lens' Environment (Maybe Text)

The ID of the environment account connection that's used to provision infrastructure resources in an environment account.

environment_spec :: Lens' Environment (Maybe Text)

The environment spec.

environment_description :: Lens' Environment (Maybe Text)

The description of the environment.

environment_arn :: Lens' Environment Text

The Amazon Resource Name (ARN) of the environment.

environment_createdAt :: Lens' Environment UTCTime

The time when the environment was created.

environment_deploymentStatus :: Lens' Environment DeploymentStatus

The environment deployment status.

environment_lastDeploymentAttemptedAt :: Lens' Environment UTCTime

The time when a deployment of the environment was last attempted.

environment_lastDeploymentSucceededAt :: Lens' Environment UTCTime

The time when the environment was last deployed successfully.

environment_name :: Lens' Environment Text

The name of the environment.

environment_templateMajorVersion :: Lens' Environment Text

The ID of the major version of the environment template.

environment_templateMinorVersion :: Lens' Environment Text

The ID of the minor version of the environment template.

environment_templateName :: Lens' Environment Text

The Amazon Resource Name (ARN) of the environment template.