Amazonka.Proton.Types.EnvironmentTemplateVersion

Description

Documentation

data EnvironmentTemplateVersion

The environment template version data.

See: newEnvironmentTemplateVersion smart constructor.

Constructors

EnvironmentTemplateVersion'

Fields

schema :: Maybe (Sensitive Text)

The schema of the version of an environment template.

statusMessage :: Maybe (Sensitive Text)

The status message of the version of an environment template.

recommendedMinorVersion :: Maybe Text

The ID of the recommended minor version of the environment template.

description :: Maybe (Sensitive Text)

A description of the minor version of an environment template.

arn :: Text

The Amazon Resource Name (ARN) of the version of an environment template.

createdAt :: POSIX

The time when the version of an environment template was created.

lastModifiedAt :: POSIX

The time when the version of an environment template was last modified.

majorVersion :: Text

The ID of the latest major version that's associated with the version of an environment template.

minorVersion :: Text

The ID of the minor version of an environment template.

status :: TemplateVersionStatus

The status of the version of an environment template.

templateName :: Text

The name of the version of an environment template.

newEnvironmentTemplateVersion

Arguments

:: Text
$sel:arn:EnvironmentTemplateVersion'

-> UTCTime
$sel:createdAt:EnvironmentTemplateVersion'

-> UTCTime
$sel:lastModifiedAt:EnvironmentTemplateVersion'

-> Text
$sel:majorVersion:EnvironmentTemplateVersion'

-> Text
$sel:minorVersion:EnvironmentTemplateVersion'

-> TemplateVersionStatus
$sel:status:EnvironmentTemplateVersion'

-> Text
$sel:templateName:EnvironmentTemplateVersion'

-> EnvironmentTemplateVersion

Create a value of EnvironmentTemplateVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schema:EnvironmentTemplateVersion', environmentTemplateVersion_schema - The schema of the version of an environment template.

$sel:statusMessage:EnvironmentTemplateVersion', environmentTemplateVersion_statusMessage - The status message of the version of an environment template.

$sel:recommendedMinorVersion:EnvironmentTemplateVersion', environmentTemplateVersion_recommendedMinorVersion - The ID of the recommended minor version of the environment template.

$sel:description:EnvironmentTemplateVersion', environmentTemplateVersion_description - A description of the minor version of an environment template.

$sel:arn:EnvironmentTemplateVersion', environmentTemplateVersion_arn - The Amazon Resource Name (ARN) of the version of an environment template.

$sel:createdAt:EnvironmentTemplateVersion', environmentTemplateVersion_createdAt - The time when the version of an environment template was created.

$sel:lastModifiedAt:EnvironmentTemplateVersion', environmentTemplateVersion_lastModifiedAt - The time when the version of an environment template was last modified.

$sel:majorVersion:EnvironmentTemplateVersion', environmentTemplateVersion_majorVersion - The ID of the latest major version that's associated with the version of an environment template.

$sel:minorVersion:EnvironmentTemplateVersion', environmentTemplateVersion_minorVersion - The ID of the minor version of an environment template.

$sel:status:EnvironmentTemplateVersion', environmentTemplateVersion_status - The status of the version of an environment template.

$sel:templateName:EnvironmentTemplateVersion', environmentTemplateVersion_templateName - The name of the version of an environment template.

environmentTemplateVersion_schema :: Lens' EnvironmentTemplateVersion (Maybe Text)

The schema of the version of an environment template.

environmentTemplateVersion_statusMessage :: Lens' EnvironmentTemplateVersion (Maybe Text)

The status message of the version of an environment template.

environmentTemplateVersion_recommendedMinorVersion :: Lens' EnvironmentTemplateVersion (Maybe Text)

The ID of the recommended minor version of the environment template.

environmentTemplateVersion_description :: Lens' EnvironmentTemplateVersion (Maybe Text)

A description of the minor version of an environment template.

environmentTemplateVersion_arn :: Lens' EnvironmentTemplateVersion Text

The Amazon Resource Name (ARN) of the version of an environment template.

environmentTemplateVersion_createdAt :: Lens' EnvironmentTemplateVersion UTCTime

The time when the version of an environment template was created.

environmentTemplateVersion_lastModifiedAt :: Lens' EnvironmentTemplateVersion UTCTime

The time when the version of an environment template was last modified.

environmentTemplateVersion_majorVersion :: Lens' EnvironmentTemplateVersion Text

The ID of the latest major version that's associated with the version of an environment template.

environmentTemplateVersion_minorVersion :: Lens' EnvironmentTemplateVersion Text

The ID of the minor version of an environment template.

environmentTemplateVersion_status :: Lens' EnvironmentTemplateVersion TemplateVersionStatus

The status of the version of an environment template.

environmentTemplateVersion_templateName :: Lens' EnvironmentTemplateVersion Text

The name of the version of an environment template.