Amazonka.Proton.Types.EnvironmentTemplateVersionSummary

Description

Documentation

data EnvironmentTemplateVersionSummary

A summary of the version of an environment template detail data.

See: newEnvironmentTemplateVersionSummary smart constructor.

Constructors

EnvironmentTemplateVersionSummary'

Fields

statusMessage :: Maybe (Sensitive Text)

The status message of the version of an environment template.

recommendedMinorVersion :: Maybe Text

The ID of the recommended minor version of the environment template.

description :: Maybe (Sensitive Text)

A description of the version of an environment template.

arn :: Text

The Amazon Resource Name (ARN) of the version of an environment template.

createdAt :: POSIX

The time when the version of an environment template was created.

lastModifiedAt :: POSIX

The time when the version of an environment template was last modified.

majorVersion :: Text

The ID of the latest major version that's associated with the version of an environment template.

minorVersion :: Text

The ID of the version of an environment template.

status :: TemplateVersionStatus

The status of the version of an environment template.

templateName :: Text

The name of the version of an environment template.

newEnvironmentTemplateVersionSummary

Arguments

:: Text
$sel:arn:EnvironmentTemplateVersionSummary'

-> UTCTime
$sel:createdAt:EnvironmentTemplateVersionSummary'

-> UTCTime
$sel:lastModifiedAt:EnvironmentTemplateVersionSummary'

-> Text
$sel:majorVersion:EnvironmentTemplateVersionSummary'

-> Text
$sel:minorVersion:EnvironmentTemplateVersionSummary'

-> TemplateVersionStatus
$sel:status:EnvironmentTemplateVersionSummary'

-> Text
$sel:templateName:EnvironmentTemplateVersionSummary'

-> EnvironmentTemplateVersionSummary

Create a value of EnvironmentTemplateVersionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:EnvironmentTemplateVersionSummary', environmentTemplateVersionSummary_statusMessage - The status message of the version of an environment template.

$sel:recommendedMinorVersion:EnvironmentTemplateVersionSummary', environmentTemplateVersionSummary_recommendedMinorVersion - The ID of the recommended minor version of the environment template.

$sel:description:EnvironmentTemplateVersionSummary', environmentTemplateVersionSummary_description - A description of the version of an environment template.

$sel:arn:EnvironmentTemplateVersionSummary', environmentTemplateVersionSummary_arn - The Amazon Resource Name (ARN) of the version of an environment template.

$sel:createdAt:EnvironmentTemplateVersionSummary', environmentTemplateVersionSummary_createdAt - The time when the version of an environment template was created.

$sel:lastModifiedAt:EnvironmentTemplateVersionSummary', environmentTemplateVersionSummary_lastModifiedAt - The time when the version of an environment template was last modified.

$sel:majorVersion:EnvironmentTemplateVersionSummary', environmentTemplateVersionSummary_majorVersion - The ID of the latest major version that's associated with the version of an environment template.

$sel:minorVersion:EnvironmentTemplateVersionSummary', environmentTemplateVersionSummary_minorVersion - The ID of the version of an environment template.

$sel:status:EnvironmentTemplateVersionSummary', environmentTemplateVersionSummary_status - The status of the version of an environment template.

$sel:templateName:EnvironmentTemplateVersionSummary', environmentTemplateVersionSummary_templateName - The name of the version of an environment template.

environmentTemplateVersionSummary_statusMessage :: Lens' EnvironmentTemplateVersionSummary (Maybe Text)

The status message of the version of an environment template.

environmentTemplateVersionSummary_recommendedMinorVersion :: Lens' EnvironmentTemplateVersionSummary (Maybe Text)

The ID of the recommended minor version of the environment template.

environmentTemplateVersionSummary_description :: Lens' EnvironmentTemplateVersionSummary (Maybe Text)

A description of the version of an environment template.

environmentTemplateVersionSummary_arn :: Lens' EnvironmentTemplateVersionSummary Text

The Amazon Resource Name (ARN) of the version of an environment template.

environmentTemplateVersionSummary_createdAt :: Lens' EnvironmentTemplateVersionSummary UTCTime

The time when the version of an environment template was created.

environmentTemplateVersionSummary_lastModifiedAt :: Lens' EnvironmentTemplateVersionSummary UTCTime

The time when the version of an environment template was last modified.

environmentTemplateVersionSummary_majorVersion :: Lens' EnvironmentTemplateVersionSummary Text

The ID of the latest major version that's associated with the version of an environment template.

environmentTemplateVersionSummary_minorVersion :: Lens' EnvironmentTemplateVersionSummary Text

The ID of the version of an environment template.

environmentTemplateVersionSummary_status :: Lens' EnvironmentTemplateVersionSummary TemplateVersionStatus

The status of the version of an environment template.

environmentTemplateVersionSummary_templateName :: Lens' EnvironmentTemplateVersionSummary Text

The name of the version of an environment template.