Amazonka.Proton.Types.Service

Description

Documentation

data Service

The service detail data.

See: newService smart constructor.

Constructors

Service'

Fields

branchName :: Maybe Text

The name of the code repository branch that holds the code that's deployed in AWS Proton.

statusMessage :: Maybe (Sensitive Text)

A service status message.

repositoryId :: Maybe Text

The ID of the code repository.

pipeline :: Maybe ServicePipeline

The service pipeline detail data.

description :: Maybe (Sensitive Text)

A description of a service.

repositoryConnectionArn :: Maybe Text

The Amazon Resource Name (ARN) of the repository connection. For more information, see Set up a repository connection in the AWS Proton Administrator Guide and Setting up with AWS Proton in the AWS Proton User Guide.

arn :: Text

The Amazon Resource Name (ARN) of the service.

createdAt :: POSIX

The time when the service was created.

lastModifiedAt :: POSIX

The time when the service was last modified.

name :: Text

The name of the service.

spec :: Sensitive Text

The formatted specification that defines the service.

status :: ServiceStatus

The status of the service.

templateName :: Text

The name of the service template.

newService

Arguments

:: Text
$sel:arn:Service'

-> UTCTime
$sel:createdAt:Service'

-> UTCTime
$sel:lastModifiedAt:Service'

-> Text
$sel:name:Service'

-> Text
$sel:spec:Service'

-> ServiceStatus
$sel:status:Service'

-> Text
$sel:templateName:Service'

-> Service

Create a value of Service with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:branchName:Service', service_branchName - The name of the code repository branch that holds the code that's deployed in AWS Proton.

$sel:statusMessage:Service', service_statusMessage - A service status message.

$sel:repositoryId:Service', service_repositoryId - The ID of the code repository.

$sel:pipeline:Service', service_pipeline - The service pipeline detail data.

$sel:description:Service', service_description - A description of a service.

$sel:repositoryConnectionArn:Service', service_repositoryConnectionArn - The Amazon Resource Name (ARN) of the repository connection. For more information, see Set up a repository connection in the AWS Proton Administrator Guide and Setting up with AWS Proton in the AWS Proton User Guide.

$sel:arn:Service', service_arn - The Amazon Resource Name (ARN) of the service.

$sel:createdAt:Service', service_createdAt - The time when the service was created.

$sel:lastModifiedAt:Service', service_lastModifiedAt - The time when the service was last modified.

$sel:name:Service', service_name - The name of the service.

$sel:spec:Service', service_spec - The formatted specification that defines the service.

$sel:status:Service', service_status - The status of the service.

$sel:templateName:Service', service_templateName - The name of the service template.

service_branchName :: Lens' Service (Maybe Text)

The name of the code repository branch that holds the code that's deployed in AWS Proton.

service_statusMessage :: Lens' Service (Maybe Text)

A service status message.

service_repositoryId :: Lens' Service (Maybe Text)

The ID of the code repository.

service_pipeline :: Lens' Service (Maybe ServicePipeline)

The service pipeline detail data.

service_description :: Lens' Service (Maybe Text)

A description of a service.

service_repositoryConnectionArn :: Lens' Service (Maybe Text)

The Amazon Resource Name (ARN) of the repository connection. For more information, see Set up a repository connection in the AWS Proton Administrator Guide and Setting up with AWS Proton in the AWS Proton User Guide.

service_arn :: Lens' Service Text

The Amazon Resource Name (ARN) of the service.

service_createdAt :: Lens' Service UTCTime

The time when the service was created.

service_lastModifiedAt :: Lens' Service UTCTime

The time when the service was last modified.

service_name :: Lens' Service Text

The name of the service.

service_spec :: Lens' Service Text

The formatted specification that defines the service.

service_status :: Lens' Service ServiceStatus

The status of the service.

service_templateName :: Lens' Service Text

The name of the service template.