Amazonka.Proton.Types.ServiceInstanceSummary

Description

Documentation

data ServiceInstanceSummary

A summary of the service instance detail data.

See: newServiceInstanceSummary smart constructor.

Constructors

ServiceInstanceSummary'

Fields

deploymentStatusMessage :: Maybe (Sensitive Text)

A service instance deployment status message.

arn :: Text

The Amazon Resource Name (ARN) of the service instance.

createdAt :: POSIX

The time when the service instance was created.

deploymentStatus :: DeploymentStatus

The service instance deployment status.

environmentName :: Text

The name of the environment that the service instance was deployed into.

lastDeploymentAttemptedAt :: POSIX

The time when a deployment of the service was last attempted.

lastDeploymentSucceededAt :: POSIX

The time when the service was last deployed successfully.

name :: Text

The name of the service instance.

serviceName :: Text

The name of the service that the service instance belongs to.

templateMajorVersion :: Text

The ID of the major version of a service template.

templateMinorVersion :: Text

The ID of the minor version of a service template.

templateName :: Text

The name of the service template.

newServiceInstanceSummary

Arguments

:: Text
$sel:arn:ServiceInstanceSummary'

-> UTCTime
$sel:createdAt:ServiceInstanceSummary'

-> DeploymentStatus
$sel:deploymentStatus:ServiceInstanceSummary'

-> Text
$sel:environmentName:ServiceInstanceSummary'

-> UTCTime
$sel:lastDeploymentAttemptedAt:ServiceInstanceSummary'

-> UTCTime
$sel:lastDeploymentSucceededAt:ServiceInstanceSummary'

-> Text
$sel:name:ServiceInstanceSummary'

-> Text
$sel:serviceName:ServiceInstanceSummary'

-> Text
$sel:templateMajorVersion:ServiceInstanceSummary'

-> Text
$sel:templateMinorVersion:ServiceInstanceSummary'

-> Text
$sel:templateName:ServiceInstanceSummary'

-> ServiceInstanceSummary

Create a value of ServiceInstanceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentStatusMessage:ServiceInstanceSummary', serviceInstanceSummary_deploymentStatusMessage - A service instance deployment status message.

$sel:arn:ServiceInstanceSummary', serviceInstanceSummary_arn - The Amazon Resource Name (ARN) of the service instance.

$sel:createdAt:ServiceInstanceSummary', serviceInstanceSummary_createdAt - The time when the service instance was created.

$sel:deploymentStatus:ServiceInstanceSummary', serviceInstanceSummary_deploymentStatus - The service instance deployment status.

$sel:environmentName:ServiceInstanceSummary', serviceInstanceSummary_environmentName - The name of the environment that the service instance was deployed into.

$sel:lastDeploymentAttemptedAt:ServiceInstanceSummary', serviceInstanceSummary_lastDeploymentAttemptedAt - The time when a deployment of the service was last attempted.

$sel:lastDeploymentSucceededAt:ServiceInstanceSummary', serviceInstanceSummary_lastDeploymentSucceededAt - The time when the service was last deployed successfully.

$sel:name:ServiceInstanceSummary', serviceInstanceSummary_name - The name of the service instance.

$sel:serviceName:ServiceInstanceSummary', serviceInstanceSummary_serviceName - The name of the service that the service instance belongs to.

$sel:templateMajorVersion:ServiceInstanceSummary', serviceInstanceSummary_templateMajorVersion - The ID of the major version of a service template.

$sel:templateMinorVersion:ServiceInstanceSummary', serviceInstanceSummary_templateMinorVersion - The ID of the minor version of a service template.

$sel:templateName:ServiceInstanceSummary', serviceInstanceSummary_templateName - The name of the service template.

serviceInstanceSummary_deploymentStatusMessage :: Lens' ServiceInstanceSummary (Maybe Text)

A service instance deployment status message.

serviceInstanceSummary_arn :: Lens' ServiceInstanceSummary Text

The Amazon Resource Name (ARN) of the service instance.

serviceInstanceSummary_createdAt :: Lens' ServiceInstanceSummary UTCTime

The time when the service instance was created.

serviceInstanceSummary_deploymentStatus :: Lens' ServiceInstanceSummary DeploymentStatus

The service instance deployment status.

serviceInstanceSummary_environmentName :: Lens' ServiceInstanceSummary Text

The name of the environment that the service instance was deployed into.

serviceInstanceSummary_lastDeploymentAttemptedAt :: Lens' ServiceInstanceSummary UTCTime

The time when a deployment of the service was last attempted.

serviceInstanceSummary_lastDeploymentSucceededAt :: Lens' ServiceInstanceSummary UTCTime

The time when the service was last deployed successfully.

serviceInstanceSummary_name :: Lens' ServiceInstanceSummary Text

The name of the service instance.

serviceInstanceSummary_serviceName :: Lens' ServiceInstanceSummary Text

The name of the service that the service instance belongs to.

serviceInstanceSummary_templateMajorVersion :: Lens' ServiceInstanceSummary Text

The ID of the major version of a service template.

serviceInstanceSummary_templateMinorVersion :: Lens' ServiceInstanceSummary Text

The ID of the minor version of a service template.

serviceInstanceSummary_templateName :: Lens' ServiceInstanceSummary Text

The name of the service template.