Amazonka.Proton.Types.ServiceTemplate

Description

Documentation

data ServiceTemplate

The service template detail data.

See: newServiceTemplate smart constructor.

Constructors

ServiceTemplate'

Fields

recommendedVersion :: Maybe Text

The ID of the recommended version of the service template.

displayName :: Maybe (Sensitive Text)

The service template name as displayed in the developer interface.

encryptionKey :: Maybe Text

The customer provided service template encryption key that's used to encrypt data.

pipelineProvisioning :: Maybe Provisioning

If pipelineProvisioning is true, a service pipeline is included in the service template. Otherwise, a service pipeline isn't included in the service template.

description :: Maybe (Sensitive Text)

A description of the service template.

arn :: Text

The Amazon Resource Name (ARN) of the service template.

createdAt :: POSIX

The time when the service template was created.

lastModifiedAt :: POSIX

The time when the service template was last modified.

name :: Text

The name of the service template.

newServiceTemplate

Arguments

:: Text
$sel:arn:ServiceTemplate'

-> UTCTime
$sel:createdAt:ServiceTemplate'

-> UTCTime
$sel:lastModifiedAt:ServiceTemplate'

-> Text
$sel:name:ServiceTemplate'

-> ServiceTemplate

Create a value of ServiceTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommendedVersion:ServiceTemplate', serviceTemplate_recommendedVersion - The ID of the recommended version of the service template.

$sel:displayName:ServiceTemplate', serviceTemplate_displayName - The service template name as displayed in the developer interface.

$sel:encryptionKey:ServiceTemplate', serviceTemplate_encryptionKey - The customer provided service template encryption key that's used to encrypt data.

$sel:pipelineProvisioning:ServiceTemplate', serviceTemplate_pipelineProvisioning - If pipelineProvisioning is true, a service pipeline is included in the service template. Otherwise, a service pipeline isn't included in the service template.

$sel:description:ServiceTemplate', serviceTemplate_description - A description of the service template.

$sel:arn:ServiceTemplate', serviceTemplate_arn - The Amazon Resource Name (ARN) of the service template.

$sel:createdAt:ServiceTemplate', serviceTemplate_createdAt - The time when the service template was created.

$sel:lastModifiedAt:ServiceTemplate', serviceTemplate_lastModifiedAt - The time when the service template was last modified.

$sel:name:ServiceTemplate', serviceTemplate_name - The name of the service template.

serviceTemplate_recommendedVersion :: Lens' ServiceTemplate (Maybe Text)

The ID of the recommended version of the service template.

serviceTemplate_displayName :: Lens' ServiceTemplate (Maybe Text)

The service template name as displayed in the developer interface.

serviceTemplate_encryptionKey :: Lens' ServiceTemplate (Maybe Text)

The customer provided service template encryption key that's used to encrypt data.

serviceTemplate_pipelineProvisioning :: Lens' ServiceTemplate (Maybe Provisioning)

If pipelineProvisioning is true, a service pipeline is included in the service template. Otherwise, a service pipeline isn't included in the service template.

serviceTemplate_description :: Lens' ServiceTemplate (Maybe Text)

A description of the service template.

serviceTemplate_arn :: Lens' ServiceTemplate Text

The Amazon Resource Name (ARN) of the service template.

serviceTemplate_createdAt :: Lens' ServiceTemplate UTCTime

The time when the service template was created.

serviceTemplate_lastModifiedAt :: Lens' ServiceTemplate UTCTime

The time when the service template was last modified.

serviceTemplate_name :: Lens' ServiceTemplate Text

The name of the service template.