Amazonka.Proton.UpdateEnvironmentTemplateVersion

Description

Update a major or minor version of an environment template.

Creating a Request

data UpdateEnvironmentTemplateVersion

See: newUpdateEnvironmentTemplateVersion smart constructor.

Constructors

UpdateEnvironmentTemplateVersion'

Fields

status :: Maybe TemplateVersionStatus

The status of the environment template minor version to update.

description :: Maybe (Sensitive Text)

A description of environment template version to update.

majorVersion :: Text

To update a major version of an environment template, include majorVersion.

minorVersion :: Text

To update a minor version of an environment template, include minorVersion.

templateName :: Text

The name of the environment template.

newUpdateEnvironmentTemplateVersion

Arguments

:: Text
$sel:majorVersion:UpdateEnvironmentTemplateVersion'

-> Text
$sel:minorVersion:UpdateEnvironmentTemplateVersion'

-> Text
$sel:templateName:UpdateEnvironmentTemplateVersion'

-> UpdateEnvironmentTemplateVersion

Create a value of UpdateEnvironmentTemplateVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateEnvironmentTemplateVersion', updateEnvironmentTemplateVersion_status - The status of the environment template minor version to update.

$sel:description:UpdateEnvironmentTemplateVersion', updateEnvironmentTemplateVersion_description - A description of environment template version to update.

$sel:majorVersion:UpdateEnvironmentTemplateVersion', updateEnvironmentTemplateVersion_majorVersion - To update a major version of an environment template, include majorVersion.

$sel:minorVersion:UpdateEnvironmentTemplateVersion', updateEnvironmentTemplateVersion_minorVersion - To update a minor version of an environment template, include minorVersion.

$sel:templateName:UpdateEnvironmentTemplateVersion', updateEnvironmentTemplateVersion_templateName - The name of the environment template.

Request Lenses

updateEnvironmentTemplateVersion_status :: Lens' UpdateEnvironmentTemplateVersion (Maybe TemplateVersionStatus)

The status of the environment template minor version to update.

updateEnvironmentTemplateVersion_description :: Lens' UpdateEnvironmentTemplateVersion (Maybe Text)

A description of environment template version to update.

updateEnvironmentTemplateVersion_majorVersion :: Lens' UpdateEnvironmentTemplateVersion Text

To update a major version of an environment template, include majorVersion.

updateEnvironmentTemplateVersion_minorVersion :: Lens' UpdateEnvironmentTemplateVersion Text

To update a minor version of an environment template, include minorVersion.

updateEnvironmentTemplateVersion_templateName :: Lens' UpdateEnvironmentTemplateVersion Text

The name of the environment template.

Destructuring the Response

data UpdateEnvironmentTemplateVersionResponse

See: newUpdateEnvironmentTemplateVersionResponse smart constructor.

Constructors

UpdateEnvironmentTemplateVersionResponse'

Fields

httpStatus :: Int

The response's http status code.

environmentTemplateVersion :: EnvironmentTemplateVersion

The environment template version detail data that's returned by AWS Proton.

newUpdateEnvironmentTemplateVersionResponse

Arguments

:: Int
$sel:httpStatus:UpdateEnvironmentTemplateVersionResponse'

-> EnvironmentTemplateVersion
$sel:environmentTemplateVersion:UpdateEnvironmentTemplateVersionResponse'

-> UpdateEnvironmentTemplateVersionResponse

Create a value of UpdateEnvironmentTemplateVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateEnvironmentTemplateVersionResponse', updateEnvironmentTemplateVersionResponse_httpStatus - The response's http status code.

$sel:environmentTemplateVersion:UpdateEnvironmentTemplateVersionResponse', updateEnvironmentTemplateVersionResponse_environmentTemplateVersion - The environment template version detail data that's returned by AWS Proton.

Response Lenses

updateEnvironmentTemplateVersionResponse_httpStatus :: Lens' UpdateEnvironmentTemplateVersionResponse Int

The response's http status code.

updateEnvironmentTemplateVersionResponse_environmentTemplateVersion :: Lens' UpdateEnvironmentTemplateVersionResponse EnvironmentTemplateVersion

The environment template version detail data that's returned by AWS Proton.