Amazonka.Proton.UpdateServiceTemplateVersion

Description

Update a major or minor version of a service template.

Creating a Request

data UpdateServiceTemplateVersion

See: newUpdateServiceTemplateVersion smart constructor.

Constructors

UpdateServiceTemplateVersion'

Fields

status :: Maybe TemplateVersionStatus

The status of the service template minor version to update.

compatibleEnvironmentTemplates :: Maybe (NonEmpty CompatibleEnvironmentTemplateInput)

An array of compatible environment names for a service template major or minor version to update.

description :: Maybe (Sensitive Text)

A description of a service template version to update.

majorVersion :: Text

To update a major version of a service template, include majorVersion.

minorVersion :: Text

To update a minor version of a service template, include minorVersion.

templateName :: Text

The name of the service template.

newUpdateServiceTemplateVersion

Arguments

:: Text
$sel:majorVersion:UpdateServiceTemplateVersion'

-> Text
$sel:minorVersion:UpdateServiceTemplateVersion'

-> Text
$sel:templateName:UpdateServiceTemplateVersion'

-> UpdateServiceTemplateVersion

Create a value of UpdateServiceTemplateVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateServiceTemplateVersion', updateServiceTemplateVersion_status - The status of the service template minor version to update.

$sel:compatibleEnvironmentTemplates:UpdateServiceTemplateVersion', updateServiceTemplateVersion_compatibleEnvironmentTemplates - An array of compatible environment names for a service template major or minor version to update.

$sel:description:UpdateServiceTemplateVersion', updateServiceTemplateVersion_description - A description of a service template version to update.

$sel:majorVersion:UpdateServiceTemplateVersion', updateServiceTemplateVersion_majorVersion - To update a major version of a service template, include majorVersion.

$sel:minorVersion:UpdateServiceTemplateVersion', updateServiceTemplateVersion_minorVersion - To update a minor version of a service template, include minorVersion.

$sel:templateName:UpdateServiceTemplateVersion', updateServiceTemplateVersion_templateName - The name of the service template.

Request Lenses

updateServiceTemplateVersion_status :: Lens' UpdateServiceTemplateVersion (Maybe TemplateVersionStatus)

The status of the service template minor version to update.

updateServiceTemplateVersion_compatibleEnvironmentTemplates :: Lens' UpdateServiceTemplateVersion (Maybe (NonEmpty CompatibleEnvironmentTemplateInput))

An array of compatible environment names for a service template major or minor version to update.

updateServiceTemplateVersion_description :: Lens' UpdateServiceTemplateVersion (Maybe Text)

A description of a service template version to update.

updateServiceTemplateVersion_majorVersion :: Lens' UpdateServiceTemplateVersion Text

To update a major version of a service template, include majorVersion.

updateServiceTemplateVersion_minorVersion :: Lens' UpdateServiceTemplateVersion Text

To update a minor version of a service template, include minorVersion.

updateServiceTemplateVersion_templateName :: Lens' UpdateServiceTemplateVersion Text

The name of the service template.

Destructuring the Response

data UpdateServiceTemplateVersionResponse

See: newUpdateServiceTemplateVersionResponse smart constructor.

Constructors

UpdateServiceTemplateVersionResponse'

Fields

httpStatus :: Int

The response's http status code.

serviceTemplateVersion :: ServiceTemplateVersion

The service template version detail data that's returned by AWS Proton.

newUpdateServiceTemplateVersionResponse

Arguments

:: Int
$sel:httpStatus:UpdateServiceTemplateVersionResponse'

-> ServiceTemplateVersion
$sel:serviceTemplateVersion:UpdateServiceTemplateVersionResponse'

-> UpdateServiceTemplateVersionResponse

Create a value of UpdateServiceTemplateVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateServiceTemplateVersionResponse', updateServiceTemplateVersionResponse_httpStatus - The response's http status code.

$sel:serviceTemplateVersion:UpdateServiceTemplateVersionResponse', updateServiceTemplateVersionResponse_serviceTemplateVersion - The service template version detail data that's returned by AWS Proton.

Response Lenses

updateServiceTemplateVersionResponse_httpStatus :: Lens' UpdateServiceTemplateVersionResponse Int

The response's http status code.

updateServiceTemplateVersionResponse_serviceTemplateVersion :: Lens' UpdateServiceTemplateVersionResponse ServiceTemplateVersion

The service template version detail data that's returned by AWS Proton.