Amazonka.QLDB.CancelJournalKinesisStream

Description

Ends a given Amazon QLDB journal stream. Before a stream can be canceled, its current status must be ACTIVE.

You can't restart a stream after you cancel it. Canceled QLDB stream resources are subject to a 7-day retention period, so they are automatically deleted after this limit expires.

Creating a Request

data CancelJournalKinesisStream

See: newCancelJournalKinesisStream smart constructor.

Constructors

CancelJournalKinesisStream'

Fields

ledgerName :: Text

The name of the ledger.

streamId :: Text

The UUID (represented in Base62-encoded text) of the QLDB journal stream to be canceled.

newCancelJournalKinesisStream

Arguments

:: Text
$sel:ledgerName:CancelJournalKinesisStream'

-> Text
$sel:streamId:CancelJournalKinesisStream'

-> CancelJournalKinesisStream

Create a value of CancelJournalKinesisStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ledgerName:CancelJournalKinesisStream', cancelJournalKinesisStream_ledgerName - The name of the ledger.

$sel:streamId:CancelJournalKinesisStream', cancelJournalKinesisStream_streamId - The UUID (represented in Base62-encoded text) of the QLDB journal stream to be canceled.

Request Lenses

cancelJournalKinesisStream_ledgerName :: Lens' CancelJournalKinesisStream Text

The name of the ledger.

cancelJournalKinesisStream_streamId :: Lens' CancelJournalKinesisStream Text

The UUID (represented in Base62-encoded text) of the QLDB journal stream to be canceled.

Destructuring the Response

data CancelJournalKinesisStreamResponse

See: newCancelJournalKinesisStreamResponse smart constructor.

Constructors

CancelJournalKinesisStreamResponse'

Fields

streamId :: Maybe Text

The UUID (Base62-encoded text) of the canceled QLDB journal stream.

httpStatus :: Int

The response's http status code.

newCancelJournalKinesisStreamResponse

Arguments

:: Int
$sel:httpStatus:CancelJournalKinesisStreamResponse'

-> CancelJournalKinesisStreamResponse

Create a value of CancelJournalKinesisStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamId:CancelJournalKinesisStream', cancelJournalKinesisStreamResponse_streamId - The UUID (Base62-encoded text) of the canceled QLDB journal stream.

$sel:httpStatus:CancelJournalKinesisStreamResponse', cancelJournalKinesisStreamResponse_httpStatus - The response's http status code.

Response Lenses

cancelJournalKinesisStreamResponse_streamId :: Lens' CancelJournalKinesisStreamResponse (Maybe Text)

The UUID (Base62-encoded text) of the canceled QLDB journal stream.

cancelJournalKinesisStreamResponse_httpStatus :: Lens' CancelJournalKinesisStreamResponse Int

The response's http status code.