Amazonka.QLDB.DescribeLedger

Description

Returns information about a ledger, including its state, permissions mode, encryption at rest settings, and when it was created.

Creating a Request

data DescribeLedger

See: newDescribeLedger smart constructor.

Constructors

DescribeLedger'

Fields

name :: Text

The name of the ledger that you want to describe.

newDescribeLedger

Arguments

:: Text
$sel:name:DescribeLedger'

-> DescribeLedger

Create a value of DescribeLedger with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeLedger', describeLedger_name - The name of the ledger that you want to describe.

Request Lenses

describeLedger_name :: Lens' DescribeLedger Text

The name of the ledger that you want to describe.

Destructuring the Response

data DescribeLedgerResponse

See: newDescribeLedgerResponse smart constructor.

Constructors

DescribeLedgerResponse'

Fields

state :: Maybe LedgerState

The current status of the ledger.

deletionProtection :: Maybe Bool

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the ledger.

encryptionDescription :: Maybe LedgerEncryptionDescription

Information about the encryption of data at rest in the ledger. This includes the current status, the KMS key, and when the key became inaccessible (in the case of an error).

name :: Maybe Text

The name of the ledger.

creationDateTime :: Maybe POSIX

The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

permissionsMode :: Maybe PermissionsMode

The permissions mode of the ledger.

httpStatus :: Int

The response's http status code.

newDescribeLedgerResponse

Arguments

:: Int
$sel:httpStatus:DescribeLedgerResponse'

-> DescribeLedgerResponse

Create a value of DescribeLedgerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:DescribeLedgerResponse', describeLedgerResponse_state - The current status of the ledger.

$sel:deletionProtection:DescribeLedgerResponse', describeLedgerResponse_deletionProtection - The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

$sel:arn:DescribeLedgerResponse', describeLedgerResponse_arn - The Amazon Resource Name (ARN) for the ledger.

$sel:encryptionDescription:DescribeLedgerResponse', describeLedgerResponse_encryptionDescription - Information about the encryption of data at rest in the ledger. This includes the current status, the KMS key, and when the key became inaccessible (in the case of an error).

$sel:name:DescribeLedger', describeLedgerResponse_name - The name of the ledger.

$sel:creationDateTime:DescribeLedgerResponse', describeLedgerResponse_creationDateTime - The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

$sel:permissionsMode:DescribeLedgerResponse', describeLedgerResponse_permissionsMode - The permissions mode of the ledger.

$sel:httpStatus:DescribeLedgerResponse', describeLedgerResponse_httpStatus - The response's http status code.

Response Lenses

describeLedgerResponse_state :: Lens' DescribeLedgerResponse (Maybe LedgerState)

The current status of the ledger.

describeLedgerResponse_deletionProtection :: Lens' DescribeLedgerResponse (Maybe Bool)

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

describeLedgerResponse_arn :: Lens' DescribeLedgerResponse (Maybe Text)

The Amazon Resource Name (ARN) for the ledger.

describeLedgerResponse_encryptionDescription :: Lens' DescribeLedgerResponse (Maybe LedgerEncryptionDescription)

Information about the encryption of data at rest in the ledger. This includes the current status, the KMS key, and when the key became inaccessible (in the case of an error).

describeLedgerResponse_name :: Lens' DescribeLedgerResponse (Maybe Text)

The name of the ledger.

describeLedgerResponse_creationDateTime :: Lens' DescribeLedgerResponse (Maybe UTCTime)

The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

describeLedgerResponse_permissionsMode :: Lens' DescribeLedgerResponse (Maybe PermissionsMode)

The permissions mode of the ledger.

describeLedgerResponse_httpStatus :: Lens' DescribeLedgerResponse Int

The response's http status code.