Amazonka.QLDB.Types.JournalS3ExportDescription

Description

Documentation

data JournalS3ExportDescription

Information about a journal export job, including the ledger name, export ID, creation time, current status, and the parameters of the original export creation request.

See: newJournalS3ExportDescription smart constructor.

Constructors

JournalS3ExportDescription'

Fields

ledgerName :: Text

The name of the ledger.

exportId :: Text

The UUID (represented in Base62-encoded text) of the journal export job.

exportCreationTime :: POSIX

The date and time, in epoch time format, when the export job was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

status :: ExportStatus

The current state of the journal export job.

inclusiveStartTime :: POSIX

The inclusive start date and time for the range of journal contents that are specified in the original export request.

exclusiveEndTime :: POSIX

The exclusive end date and time for the range of journal contents that are specified in the original export request.

s3ExportConfiguration :: S3ExportConfiguration

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal export job to do the following:

Write objects into your Amazon Simple Storage Service (Amazon S3) bucket.

(Optional) Use your customer master key (CMK) in Key Management Service (KMS) for server-side encryption of your exported data.

newJournalS3ExportDescription

Arguments

:: Text
$sel:ledgerName:JournalS3ExportDescription'

-> Text
$sel:exportId:JournalS3ExportDescription'

-> UTCTime
$sel:exportCreationTime:JournalS3ExportDescription'

-> ExportStatus
$sel:status:JournalS3ExportDescription'

-> UTCTime
$sel:inclusiveStartTime:JournalS3ExportDescription'

-> UTCTime
$sel:exclusiveEndTime:JournalS3ExportDescription'

-> S3ExportConfiguration
$sel:s3ExportConfiguration:JournalS3ExportDescription'

-> Text
$sel:roleArn:JournalS3ExportDescription'

-> JournalS3ExportDescription

Create a value of JournalS3ExportDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ledgerName:JournalS3ExportDescription', journalS3ExportDescription_ledgerName - The name of the ledger.

$sel:exportId:JournalS3ExportDescription', journalS3ExportDescription_exportId - The UUID (represented in Base62-encoded text) of the journal export job.

$sel:exportCreationTime:JournalS3ExportDescription', journalS3ExportDescription_exportCreationTime - The date and time, in epoch time format, when the export job was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

$sel:status:JournalS3ExportDescription', journalS3ExportDescription_status - The current state of the journal export job.

$sel:inclusiveStartTime:JournalS3ExportDescription', journalS3ExportDescription_inclusiveStartTime - The inclusive start date and time for the range of journal contents that are specified in the original export request.

$sel:exclusiveEndTime:JournalS3ExportDescription', journalS3ExportDescription_exclusiveEndTime - The exclusive end date and time for the range of journal contents that are specified in the original export request.

$sel:s3ExportConfiguration:JournalS3ExportDescription', journalS3ExportDescription_s3ExportConfiguration - Undocumented member.

$sel:roleArn:JournalS3ExportDescription', journalS3ExportDescription_roleArn - The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal export job to do the following:

Write objects into your Amazon Simple Storage Service (Amazon S3) bucket.

(Optional) Use your customer master key (CMK) in Key Management Service (KMS) for server-side encryption of your exported data.

journalS3ExportDescription_ledgerName :: Lens' JournalS3ExportDescription Text

The name of the ledger.

journalS3ExportDescription_exportId :: Lens' JournalS3ExportDescription Text

The UUID (represented in Base62-encoded text) of the journal export job.

journalS3ExportDescription_exportCreationTime :: Lens' JournalS3ExportDescription UTCTime

The date and time, in epoch time format, when the export job was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

journalS3ExportDescription_status :: Lens' JournalS3ExportDescription ExportStatus

The current state of the journal export job.

journalS3ExportDescription_inclusiveStartTime :: Lens' JournalS3ExportDescription UTCTime

The inclusive start date and time for the range of journal contents that are specified in the original export request.

journalS3ExportDescription_exclusiveEndTime :: Lens' JournalS3ExportDescription UTCTime

The exclusive end date and time for the range of journal contents that are specified in the original export request.

journalS3ExportDescription_s3ExportConfiguration :: Lens' JournalS3ExportDescription S3ExportConfiguration

Undocumented member.

journalS3ExportDescription_roleArn :: Lens' JournalS3ExportDescription Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal export job to do the following:

Write objects into your Amazon Simple Storage Service (Amazon S3) bucket.

(Optional) Use your customer master key (CMK) in Key Management Service (KMS) for server-side encryption of your exported data.