Amazonka.QLDB.Types.LedgerEncryptionDescription

Description

Documentation

data LedgerEncryptionDescription

Information about the encryption of data at rest in an Amazon QLDB ledger. This includes the current status, the key in Key Management Service (KMS), and when the key became inaccessible (in the case of an error).

For more information, see Encryption at rest in the Amazon QLDB Developer Guide.

See: newLedgerEncryptionDescription smart constructor.

Constructors

LedgerEncryptionDescription'

Fields

inaccessibleKmsKeyDateTime :: Maybe POSIX

The date and time, in epoch time format, when the KMS key first became inaccessible, in the case of an error. (Epoch time format is the number of seconds that have elapsed since 12:00:00 AM January 1, 1970 UTC.)

This parameter is undefined if the KMS key is accessible.

kmsKeyArn :: Text

The Amazon Resource Name (ARN) of the customer managed KMS key that the ledger uses for encryption at rest. If this parameter is undefined, the ledger uses an Amazon Web Services owned KMS key for encryption.

encryptionStatus :: EncryptionStatus

The current state of encryption at rest for the ledger. This can be one of the following values:

ENABLED: Encryption is fully enabled using the specified key.

UPDATING: The ledger is actively processing the specified key change.

Key changes in QLDB are asynchronous. The ledger is fully accessible without any performance impact while the key change is being processed. The amount of time it takes to update a key varies depending on the ledger size.

KMS_KEY_INACCESSIBLE: The specified customer managed KMS key is not accessible, and the ledger is impaired. Either the key was disabled or deleted, or the grants on the key were revoked. When a ledger is impaired, it is not accessible and does not accept any read or write requests.

An impaired ledger automatically returns to an active state after you restore the grants on the key, or re-enable the key that was disabled. However, deleting a customer managed KMS key is irreversible. After a key is deleted, you can no longer access the ledgers that are protected with that key, and the data becomes unrecoverable permanently.

newLedgerEncryptionDescription

Arguments

:: Text
$sel:kmsKeyArn:LedgerEncryptionDescription'

-> EncryptionStatus
$sel:encryptionStatus:LedgerEncryptionDescription'

-> LedgerEncryptionDescription

Create a value of LedgerEncryptionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inaccessibleKmsKeyDateTime:LedgerEncryptionDescription', ledgerEncryptionDescription_inaccessibleKmsKeyDateTime - The date and time, in epoch time format, when the KMS key first became inaccessible, in the case of an error. (Epoch time format is the number of seconds that have elapsed since 12:00:00 AM January 1, 1970 UTC.)

This parameter is undefined if the KMS key is accessible.

$sel:kmsKeyArn:LedgerEncryptionDescription', ledgerEncryptionDescription_kmsKeyArn - The Amazon Resource Name (ARN) of the customer managed KMS key that the ledger uses for encryption at rest. If this parameter is undefined, the ledger uses an Amazon Web Services owned KMS key for encryption.

$sel:encryptionStatus:LedgerEncryptionDescription', ledgerEncryptionDescription_encryptionStatus - The current state of encryption at rest for the ledger. This can be one of the following values:

ENABLED: Encryption is fully enabled using the specified key.

UPDATING: The ledger is actively processing the specified key change.

Key changes in QLDB are asynchronous. The ledger is fully accessible without any performance impact while the key change is being processed. The amount of time it takes to update a key varies depending on the ledger size.

KMS_KEY_INACCESSIBLE: The specified customer managed KMS key is not accessible, and the ledger is impaired. Either the key was disabled or deleted, or the grants on the key were revoked. When a ledger is impaired, it is not accessible and does not accept any read or write requests.

An impaired ledger automatically returns to an active state after you restore the grants on the key, or re-enable the key that was disabled. However, deleting a customer managed KMS key is irreversible. After a key is deleted, you can no longer access the ledgers that are protected with that key, and the data becomes unrecoverable permanently.

ledgerEncryptionDescription_inaccessibleKmsKeyDateTime :: Lens' LedgerEncryptionDescription (Maybe UTCTime)

The date and time, in epoch time format, when the KMS key first became inaccessible, in the case of an error. (Epoch time format is the number of seconds that have elapsed since 12:00:00 AM January 1, 1970 UTC.)

This parameter is undefined if the KMS key is accessible.

ledgerEncryptionDescription_kmsKeyArn :: Lens' LedgerEncryptionDescription Text

The Amazon Resource Name (ARN) of the customer managed KMS key that the ledger uses for encryption at rest. If this parameter is undefined, the ledger uses an Amazon Web Services owned KMS key for encryption.

ledgerEncryptionDescription_encryptionStatus :: Lens' LedgerEncryptionDescription EncryptionStatus

The current state of encryption at rest for the ledger. This can be one of the following values:

ENABLED: Encryption is fully enabled using the specified key.

UPDATING: The ledger is actively processing the specified key change.

Key changes in QLDB are asynchronous. The ledger is fully accessible without any performance impact while the key change is being processed. The amount of time it takes to update a key varies depending on the ledger size.

KMS_KEY_INACCESSIBLE: The specified customer managed KMS key is not accessible, and the ledger is impaired. Either the key was disabled or deleted, or the grants on the key were revoked. When a ledger is impaired, it is not accessible and does not accept any read or write requests.

An impaired ledger automatically returns to an active state after you restore the grants on the key, or re-enable the key that was disabled. However, deleting a customer managed KMS key is irreversible. After a key is deleted, you can no longer access the ledgers that are protected with that key, and the data becomes unrecoverable permanently.