Amazonka.QLDBSession.SendCommand

Description

Sends a command to an Amazon QLDB ledger.

Instead of interacting directly with this API, we recommend using the QLDB driver or the QLDB shell to execute data transactions on a ledger.

If you are working with an AWS SDK, use the QLDB driver. The driver provides a high-level abstraction layer above this QLDB Session data plane and manages SendCommand API calls for you. For information and a list of supported programming languages, see Getting started with the driver in the Amazon QLDB Developer Guide.

If you are working with the AWS Command Line Interface (AWS CLI), use the QLDB shell. The shell is a command line interface that uses the QLDB driver to interact with a ledger. For information, see Accessing Amazon QLDB using the QLDB shell.

Creating a Request

data SendCommand

See: newSendCommand smart constructor.

Constructors

SendCommand'

Fields

fetchPage :: Maybe FetchPageRequest

Command to fetch a page.

sessionToken :: Maybe Text

Specifies the session token for the current command. A session token is constant throughout the life of the session.

To obtain a session token, run the StartSession command. This SessionToken is required for every subsequent command that is issued during the current session.

abortTransaction :: Maybe AbortTransactionRequest

Command to abort the current transaction.

endSession :: Maybe EndSessionRequest

Command to end the current session.

startTransaction :: Maybe StartTransactionRequest

Command to start a new transaction.

executeStatement :: Maybe ExecuteStatementRequest

Command to execute a statement in the specified transaction.

commitTransaction :: Maybe CommitTransactionRequest

Command to commit the specified transaction.

startSession :: Maybe StartSessionRequest

Command to start a new session. A session token is obtained as part of the response.

newSendCommand :: SendCommand

Create a value of SendCommand with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fetchPage:SendCommand', sendCommand_fetchPage - Command to fetch a page.

$sel:sessionToken:SendCommand', sendCommand_sessionToken - Specifies the session token for the current command. A session token is constant throughout the life of the session.

To obtain a session token, run the StartSession command. This SessionToken is required for every subsequent command that is issued during the current session.

$sel:abortTransaction:SendCommand', sendCommand_abortTransaction - Command to abort the current transaction.

$sel:endSession:SendCommand', sendCommand_endSession - Command to end the current session.

$sel:startTransaction:SendCommand', sendCommand_startTransaction - Command to start a new transaction.

$sel:executeStatement:SendCommand', sendCommand_executeStatement - Command to execute a statement in the specified transaction.

$sel:commitTransaction:SendCommand', sendCommand_commitTransaction - Command to commit the specified transaction.

$sel:startSession:SendCommand', sendCommand_startSession - Command to start a new session. A session token is obtained as part of the response.

Request Lenses

sendCommand_fetchPage :: Lens' SendCommand (Maybe FetchPageRequest)

Command to fetch a page.

sendCommand_sessionToken :: Lens' SendCommand (Maybe Text)

Specifies the session token for the current command. A session token is constant throughout the life of the session.

To obtain a session token, run the StartSession command. This SessionToken is required for every subsequent command that is issued during the current session.

sendCommand_abortTransaction :: Lens' SendCommand (Maybe AbortTransactionRequest)

Command to abort the current transaction.

sendCommand_endSession :: Lens' SendCommand (Maybe EndSessionRequest)

Command to end the current session.

sendCommand_startTransaction :: Lens' SendCommand (Maybe StartTransactionRequest)

Command to start a new transaction.

sendCommand_executeStatement :: Lens' SendCommand (Maybe ExecuteStatementRequest)

Command to execute a statement in the specified transaction.

sendCommand_commitTransaction :: Lens' SendCommand (Maybe CommitTransactionRequest)

Command to commit the specified transaction.

sendCommand_startSession :: Lens' SendCommand (Maybe StartSessionRequest)

Command to start a new session. A session token is obtained as part of the response.

Destructuring the Response

data SendCommandResponse

See: newSendCommandResponse smart constructor.

Constructors

SendCommandResponse'

Fields

fetchPage :: Maybe FetchPageResult

Contains the details of the fetched page.

abortTransaction :: Maybe AbortTransactionResult

Contains the details of the aborted transaction.

endSession :: Maybe EndSessionResult

Contains the details of the ended session.

startTransaction :: Maybe StartTransactionResult

Contains the details of the started transaction.

executeStatement :: Maybe ExecuteStatementResult

Contains the details of the executed statement.

commitTransaction :: Maybe CommitTransactionResult

Contains the details of the committed transaction.

startSession :: Maybe StartSessionResult

Contains the details of the started session that includes a session token. This SessionToken is required for every subsequent command that is issued during the current session.

httpStatus :: Int

The response's http status code.

newSendCommandResponse

Arguments

:: Int
$sel:httpStatus:SendCommandResponse'

-> SendCommandResponse

Create a value of SendCommandResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fetchPage:SendCommand', sendCommandResponse_fetchPage - Contains the details of the fetched page.

$sel:abortTransaction:SendCommand', sendCommandResponse_abortTransaction - Contains the details of the aborted transaction.

$sel:endSession:SendCommand', sendCommandResponse_endSession - Contains the details of the ended session.

$sel:startTransaction:SendCommand', sendCommandResponse_startTransaction - Contains the details of the started transaction.

$sel:executeStatement:SendCommand', sendCommandResponse_executeStatement - Contains the details of the executed statement.

$sel:commitTransaction:SendCommand', sendCommandResponse_commitTransaction - Contains the details of the committed transaction.

$sel:startSession:SendCommand', sendCommandResponse_startSession - Contains the details of the started session that includes a session token. This SessionToken is required for every subsequent command that is issued during the current session.

$sel:httpStatus:SendCommandResponse', sendCommandResponse_httpStatus - The response's http status code.

Response Lenses

sendCommandResponse_fetchPage :: Lens' SendCommandResponse (Maybe FetchPageResult)

Contains the details of the fetched page.

sendCommandResponse_abortTransaction :: Lens' SendCommandResponse (Maybe AbortTransactionResult)

Contains the details of the aborted transaction.

sendCommandResponse_endSession :: Lens' SendCommandResponse (Maybe EndSessionResult)

Contains the details of the ended session.

sendCommandResponse_startTransaction :: Lens' SendCommandResponse (Maybe StartTransactionResult)

Contains the details of the started transaction.

sendCommandResponse_executeStatement :: Lens' SendCommandResponse (Maybe ExecuteStatementResult)

Contains the details of the executed statement.

sendCommandResponse_commitTransaction :: Lens' SendCommandResponse (Maybe CommitTransactionResult)

Contains the details of the committed transaction.

sendCommandResponse_startSession :: Lens' SendCommandResponse (Maybe StartSessionResult)

Contains the details of the started session that includes a session token. This SessionToken is required for every subsequent command that is issued during the current session.

sendCommandResponse_httpStatus :: Lens' SendCommandResponse Int

The response's http status code.