Amazonka.QLDBSession.Types

Description

Service Configuration

defaultService :: Service

API version 2019-07-11 of the Amazon QLDB Session SDK configuration.

Errors

_InvalidSessionException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the session doesn't exist anymore because it timed out or expired.

_CapacityExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Returned when the request exceeds the processing capacity of the ledger.

_OccConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Returned when a transaction cannot be written to the journal due to a failure in the verification phase of optimistic concurrency control (OCC).

_RateExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Returned when the rate of requests exceeds the allowed throughput.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if the request is malformed or contains an error such as an invalid parameter value or a missing required parameter.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Returned if a resource limit such as number of active sessions is exceeded.

AbortTransactionRequest

data AbortTransactionRequest

Contains the details of the transaction to abort.

See: newAbortTransactionRequest smart constructor.

Constructors

AbortTransactionRequest'

newAbortTransactionRequest :: AbortTransactionRequest

Create a value of AbortTransactionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

AbortTransactionResult

data AbortTransactionResult

Contains the details of the aborted transaction.

See: newAbortTransactionResult smart constructor.

Constructors

AbortTransactionResult'

Fields

timingInformation :: Maybe TimingInformation

Contains server-side performance information for the command.

newAbortTransactionResult :: AbortTransactionResult

Create a value of AbortTransactionResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timingInformation:AbortTransactionResult', abortTransactionResult_timingInformation - Contains server-side performance information for the command.

abortTransactionResult_timingInformation :: Lens' AbortTransactionResult (Maybe TimingInformation)

Contains server-side performance information for the command.

CommitTransactionRequest

data CommitTransactionRequest

Contains the details of the transaction to commit.

See: newCommitTransactionRequest smart constructor.

Constructors

CommitTransactionRequest'

Fields

transactionId :: Text

Specifies the transaction ID of the transaction to commit.

commitDigest :: Base64

Specifies the commit digest for the transaction to commit. For every active transaction, the commit digest must be passed. QLDB validates CommitDigest and rejects the commit with an error if the digest computed on the client does not match the digest computed by QLDB.

The purpose of the CommitDigest parameter is to ensure that QLDB commits a transaction if and only if the server has processed the exact set of statements sent by the client, in the same order that client sent them, and with no duplicates.

newCommitTransactionRequest

Arguments

:: Text
$sel:transactionId:CommitTransactionRequest'

-> ByteString
$sel:commitDigest:CommitTransactionRequest'

-> CommitTransactionRequest

Create a value of CommitTransactionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transactionId:CommitTransactionRequest', commitTransactionRequest_transactionId - Specifies the transaction ID of the transaction to commit.

$sel:commitDigest:CommitTransactionRequest', commitTransactionRequest_commitDigest - Specifies the commit digest for the transaction to commit. For every active transaction, the commit digest must be passed. QLDB validates CommitDigest and rejects the commit with an error if the digest computed on the client does not match the digest computed by QLDB.

The purpose of the CommitDigest parameter is to ensure that QLDB commits a transaction if and only if the server has processed the exact set of statements sent by the client, in the same order that client sent them, and with no duplicates.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

commitTransactionRequest_transactionId :: Lens' CommitTransactionRequest Text

Specifies the transaction ID of the transaction to commit.

commitTransactionRequest_commitDigest :: Lens' CommitTransactionRequest ByteString

Specifies the commit digest for the transaction to commit. For every active transaction, the commit digest must be passed. QLDB validates CommitDigest and rejects the commit with an error if the digest computed on the client does not match the digest computed by QLDB.

The purpose of the CommitDigest parameter is to ensure that QLDB commits a transaction if and only if the server has processed the exact set of statements sent by the client, in the same order that client sent them, and with no duplicates.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

CommitTransactionResult

data CommitTransactionResult

Contains the details of the committed transaction.

See: newCommitTransactionResult smart constructor.

Constructors

CommitTransactionResult'

Fields

timingInformation :: Maybe TimingInformation

Contains server-side performance information for the command.

consumedIOs :: Maybe IOUsage

Contains metrics about the number of I/O requests that were consumed.

commitDigest :: Maybe Base64

The commit digest of the committed transaction.

transactionId :: Maybe Text

The transaction ID of the committed transaction.

newCommitTransactionResult :: CommitTransactionResult

Create a value of CommitTransactionResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timingInformation:CommitTransactionResult', commitTransactionResult_timingInformation - Contains server-side performance information for the command.

$sel:consumedIOs:CommitTransactionResult', commitTransactionResult_consumedIOs - Contains metrics about the number of I/O requests that were consumed.

$sel:commitDigest:CommitTransactionResult', commitTransactionResult_commitDigest - The commit digest of the committed transaction.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:transactionId:CommitTransactionResult', commitTransactionResult_transactionId - The transaction ID of the committed transaction.

commitTransactionResult_timingInformation :: Lens' CommitTransactionResult (Maybe TimingInformation)

Contains server-side performance information for the command.

commitTransactionResult_consumedIOs :: Lens' CommitTransactionResult (Maybe IOUsage)

Contains metrics about the number of I/O requests that were consumed.

commitTransactionResult_commitDigest :: Lens' CommitTransactionResult (Maybe ByteString)

The commit digest of the committed transaction.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

commitTransactionResult_transactionId :: Lens' CommitTransactionResult (Maybe Text)

The transaction ID of the committed transaction.

EndSessionRequest

data EndSessionRequest

Specifies a request to end the session.

See: newEndSessionRequest smart constructor.

Constructors

EndSessionRequest'

newEndSessionRequest :: EndSessionRequest

Create a value of EndSessionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

EndSessionResult

data EndSessionResult

Contains the details of the ended session.

See: newEndSessionResult smart constructor.

Constructors

EndSessionResult'

Fields

timingInformation :: Maybe TimingInformation

Contains server-side performance information for the command.

newEndSessionResult :: EndSessionResult

Create a value of EndSessionResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timingInformation:EndSessionResult', endSessionResult_timingInformation - Contains server-side performance information for the command.

endSessionResult_timingInformation :: Lens' EndSessionResult (Maybe TimingInformation)

Contains server-side performance information for the command.

ExecuteStatementRequest

data ExecuteStatementRequest

Specifies a request to execute a statement.

See: newExecuteStatementRequest smart constructor.

Constructors

ExecuteStatementRequest'

Fields

parameters :: Maybe [ValueHolder]

Specifies the parameters for the parameterized statement in the request.

transactionId :: Text

Specifies the transaction ID of the request.

statement :: Text

Specifies the statement of the request.

newExecuteStatementRequest

Arguments

:: Text
$sel:transactionId:ExecuteStatementRequest'

-> Text
$sel:statement:ExecuteStatementRequest'

-> ExecuteStatementRequest

Create a value of ExecuteStatementRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:ExecuteStatementRequest', executeStatementRequest_parameters - Specifies the parameters for the parameterized statement in the request.

$sel:transactionId:ExecuteStatementRequest', executeStatementRequest_transactionId - Specifies the transaction ID of the request.

$sel:statement:ExecuteStatementRequest', executeStatementRequest_statement - Specifies the statement of the request.

executeStatementRequest_parameters :: Lens' ExecuteStatementRequest (Maybe [ValueHolder])

Specifies the parameters for the parameterized statement in the request.

executeStatementRequest_transactionId :: Lens' ExecuteStatementRequest Text

Specifies the transaction ID of the request.

executeStatementRequest_statement :: Lens' ExecuteStatementRequest Text

Specifies the statement of the request.

ExecuteStatementResult

data ExecuteStatementResult

Contains the details of the executed statement.

See: newExecuteStatementResult smart constructor.

Constructors

ExecuteStatementResult'

Fields

timingInformation :: Maybe TimingInformation

Contains server-side performance information for the command.

consumedIOs :: Maybe IOUsage

Contains metrics about the number of I/O requests that were consumed.

firstPage :: Maybe Page

Contains the details of the first fetched page.

newExecuteStatementResult :: ExecuteStatementResult

Create a value of ExecuteStatementResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timingInformation:ExecuteStatementResult', executeStatementResult_timingInformation - Contains server-side performance information for the command.

$sel:consumedIOs:ExecuteStatementResult', executeStatementResult_consumedIOs - Contains metrics about the number of I/O requests that were consumed.

$sel:firstPage:ExecuteStatementResult', executeStatementResult_firstPage - Contains the details of the first fetched page.

executeStatementResult_timingInformation :: Lens' ExecuteStatementResult (Maybe TimingInformation)

Contains server-side performance information for the command.

executeStatementResult_consumedIOs :: Lens' ExecuteStatementResult (Maybe IOUsage)

Contains metrics about the number of I/O requests that were consumed.

executeStatementResult_firstPage :: Lens' ExecuteStatementResult (Maybe Page)

Contains the details of the first fetched page.

FetchPageRequest

data FetchPageRequest

Specifies the details of the page to be fetched.

See: newFetchPageRequest smart constructor.

Constructors

FetchPageRequest'

Fields

transactionId :: Text

Specifies the transaction ID of the page to be fetched.

nextPageToken :: Text

Specifies the next page token of the page to be fetched.

newFetchPageRequest

Arguments

:: Text
$sel:transactionId:FetchPageRequest'

-> Text
$sel:nextPageToken:FetchPageRequest'

-> FetchPageRequest

Create a value of FetchPageRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transactionId:FetchPageRequest', fetchPageRequest_transactionId - Specifies the transaction ID of the page to be fetched.

$sel:nextPageToken:FetchPageRequest', fetchPageRequest_nextPageToken - Specifies the next page token of the page to be fetched.

fetchPageRequest_transactionId :: Lens' FetchPageRequest Text

Specifies the transaction ID of the page to be fetched.

fetchPageRequest_nextPageToken :: Lens' FetchPageRequest Text

Specifies the next page token of the page to be fetched.

FetchPageResult

data FetchPageResult

Contains the page that was fetched.

See: newFetchPageResult smart constructor.

Constructors

FetchPageResult'

Fields

timingInformation :: Maybe TimingInformation

Contains server-side performance information for the command.

consumedIOs :: Maybe IOUsage

Contains metrics about the number of I/O requests that were consumed.

page :: Maybe Page

Contains details of the fetched page.

newFetchPageResult :: FetchPageResult

Create a value of FetchPageResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timingInformation:FetchPageResult', fetchPageResult_timingInformation - Contains server-side performance information for the command.

$sel:consumedIOs:FetchPageResult', fetchPageResult_consumedIOs - Contains metrics about the number of I/O requests that were consumed.

$sel:page:FetchPageResult', fetchPageResult_page - Contains details of the fetched page.

fetchPageResult_timingInformation :: Lens' FetchPageResult (Maybe TimingInformation)

Contains server-side performance information for the command.

fetchPageResult_consumedIOs :: Lens' FetchPageResult (Maybe IOUsage)

Contains metrics about the number of I/O requests that were consumed.

fetchPageResult_page :: Lens' FetchPageResult (Maybe Page)

Contains details of the fetched page.

IOUsage

data IOUsage

Contains I/O usage metrics for a command that was invoked.

See: newIOUsage smart constructor.

Constructors

IOUsage'

Fields

readIOs :: Maybe Integer

The number of read I/O requests that the command made.

writeIOs :: Maybe Integer

The number of write I/O requests that the command made.

newIOUsage :: IOUsage

Create a value of IOUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readIOs:IOUsage', iOUsage_readIOs - The number of read I/O requests that the command made.

$sel:writeIOs:IOUsage', iOUsage_writeIOs - The number of write I/O requests that the command made.

iOUsage_readIOs :: Lens' IOUsage (Maybe Integer)

The number of read I/O requests that the command made.

iOUsage_writeIOs :: Lens' IOUsage (Maybe Integer)

The number of write I/O requests that the command made.

Page

data Page

Contains details of the fetched page.

See: newPage smart constructor.

Constructors

Page'

Fields

nextPageToken :: Maybe Text

The token of the next page.

values :: Maybe [ValueHolder]

A structure that contains values in multiple encoding formats.

newPage :: Page

Create a value of Page with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:Page', page_nextPageToken - The token of the next page.

$sel:values:Page', page_values - A structure that contains values in multiple encoding formats.

page_nextPageToken :: Lens' Page (Maybe Text)

The token of the next page.

page_values :: Lens' Page (Maybe [ValueHolder])

A structure that contains values in multiple encoding formats.

StartSessionRequest

data StartSessionRequest

Specifies a request to start a new session.

See: newStartSessionRequest smart constructor.

Constructors

StartSessionRequest'

Fields

ledgerName :: Text

The name of the ledger to start a new session against.

newStartSessionRequest

Arguments

:: Text
$sel:ledgerName:StartSessionRequest'

-> StartSessionRequest

Create a value of StartSessionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ledgerName:StartSessionRequest', startSessionRequest_ledgerName - The name of the ledger to start a new session against.

startSessionRequest_ledgerName :: Lens' StartSessionRequest Text

The name of the ledger to start a new session against.

StartSessionResult

data StartSessionResult

Contains the details of the started session.

See: newStartSessionResult smart constructor.

Constructors

StartSessionResult'

Fields

timingInformation :: Maybe TimingInformation

Contains server-side performance information for the command.

sessionToken :: Maybe Text

Session token of the started session. This SessionToken is required for every subsequent command that is issued during the current session.

newStartSessionResult :: StartSessionResult

Create a value of StartSessionResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timingInformation:StartSessionResult', startSessionResult_timingInformation - Contains server-side performance information for the command.

$sel:sessionToken:StartSessionResult', startSessionResult_sessionToken - Session token of the started session. This SessionToken is required for every subsequent command that is issued during the current session.

startSessionResult_timingInformation :: Lens' StartSessionResult (Maybe TimingInformation)

Contains server-side performance information for the command.

startSessionResult_sessionToken :: Lens' StartSessionResult (Maybe Text)

Session token of the started session. This SessionToken is required for every subsequent command that is issued during the current session.

StartTransactionRequest

data StartTransactionRequest

Specifies a request to start a transaction.

See: newStartTransactionRequest smart constructor.

Constructors

StartTransactionRequest'

newStartTransactionRequest :: StartTransactionRequest

Create a value of StartTransactionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

StartTransactionResult

data StartTransactionResult

Contains the details of the started transaction.

See: newStartTransactionResult smart constructor.

Constructors

StartTransactionResult'

Fields

timingInformation :: Maybe TimingInformation

Contains server-side performance information for the command.

transactionId :: Maybe Text

The transaction ID of the started transaction.

newStartTransactionResult :: StartTransactionResult

Create a value of StartTransactionResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timingInformation:StartTransactionResult', startTransactionResult_timingInformation - Contains server-side performance information for the command.

$sel:transactionId:StartTransactionResult', startTransactionResult_transactionId - The transaction ID of the started transaction.

startTransactionResult_timingInformation :: Lens' StartTransactionResult (Maybe TimingInformation)

Contains server-side performance information for the command.

startTransactionResult_transactionId :: Lens' StartTransactionResult (Maybe Text)

The transaction ID of the started transaction.

TimingInformation

data TimingInformation

Contains server-side performance information for a command. Amazon QLDB captures timing information between the times when it receives the request and when it sends the corresponding response.

See: newTimingInformation smart constructor.

Constructors

TimingInformation'

Fields

processingTimeMilliseconds :: Maybe Integer

The amount of time that QLDB spent on processing the command, measured in milliseconds.

newTimingInformation :: TimingInformation

Create a value of TimingInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:processingTimeMilliseconds:TimingInformation', timingInformation_processingTimeMilliseconds - The amount of time that QLDB spent on processing the command, measured in milliseconds.

timingInformation_processingTimeMilliseconds :: Lens' TimingInformation (Maybe Integer)

The amount of time that QLDB spent on processing the command, measured in milliseconds.

ValueHolder

data ValueHolder

A structure that can contain a value in multiple encoding formats.

See: newValueHolder smart constructor.

Constructors

ValueHolder'

Fields

ionText :: Maybe Text

An Amazon Ion plaintext value contained in a ValueHolder structure.

ionBinary :: Maybe Base64

An Amazon Ion binary value contained in a ValueHolder structure.

newValueHolder :: ValueHolder

Create a value of ValueHolder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ionText:ValueHolder', valueHolder_ionText - An Amazon Ion plaintext value contained in a ValueHolder structure.

$sel:ionBinary:ValueHolder', valueHolder_ionBinary - An Amazon Ion binary value contained in a ValueHolder structure.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

valueHolder_ionText :: Lens' ValueHolder (Maybe Text)

An Amazon Ion plaintext value contained in a ValueHolder structure.

valueHolder_ionBinary :: Lens' ValueHolder (Maybe ByteString)

An Amazon Ion binary value contained in a ValueHolder structure.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.