Amazonka.QuickSight.CreateDataSet

Description

Creates a dataset.

Creating a Request

data CreateDataSet

See: newCreateDataSet smart constructor.

Constructors

CreateDataSet'

Fields

fieldFolders :: Maybe (HashMap Text FieldFolder)

The folder that contains fields and nested subfolders for your dataset.

columnGroups :: Maybe (NonEmpty ColumnGroup)

Groupings of columns that work together in certain Amazon QuickSight features. Currently, only geospatial hierarchy is supported.

dataSetUsageConfiguration :: Maybe DataSetUsageConfiguration

rowLevelPermissionTagConfiguration :: Maybe RowLevelPermissionTagConfiguration

The configuration of tags on a dataset to set row-level security. Row-level security tags are currently supported for anonymous embedding only.

rowLevelPermissionDataSet :: Maybe RowLevelPermissionDataSet

The row-level security configuration for the data that you want to create.

permissions :: Maybe (NonEmpty ResourcePermission)

A list of resource permissions on the dataset.

columnLevelPermissionRules :: Maybe (NonEmpty ColumnLevelPermissionRule)

A set of one or more definitions of a ColumnLevelPermissionRule .

tags :: Maybe (NonEmpty Tag)

Contains a map of the key-value pairs for the resource tag or tags assigned to the dataset.

logicalTableMap :: Maybe (HashMap Text LogicalTable)

Configures the combination and transformation of the data from the physical tables.

awsAccountId :: Text

The Amazon Web Services account ID.

dataSetId :: Text

An ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

name :: Text

The display name for the dataset.

physicalTableMap :: HashMap Text PhysicalTable

Declares the physical tables that are available in the underlying data sources.

importMode :: DataSetImportMode

Indicates whether you want to import the data into SPICE.

newCreateDataSet

Arguments

:: Text
$sel:awsAccountId:CreateDataSet'

-> Text
$sel:dataSetId:CreateDataSet'

-> Text
$sel:name:CreateDataSet'

-> DataSetImportMode
$sel:importMode:CreateDataSet'

-> CreateDataSet

Create a value of CreateDataSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fieldFolders:CreateDataSet', createDataSet_fieldFolders - The folder that contains fields and nested subfolders for your dataset.

$sel:columnGroups:CreateDataSet', createDataSet_columnGroups - Groupings of columns that work together in certain Amazon QuickSight features. Currently, only geospatial hierarchy is supported.

$sel:dataSetUsageConfiguration:CreateDataSet', createDataSet_dataSetUsageConfiguration - Undocumented member.

$sel:rowLevelPermissionTagConfiguration:CreateDataSet', createDataSet_rowLevelPermissionTagConfiguration - The configuration of tags on a dataset to set row-level security. Row-level security tags are currently supported for anonymous embedding only.

$sel:rowLevelPermissionDataSet:CreateDataSet', createDataSet_rowLevelPermissionDataSet - The row-level security configuration for the data that you want to create.

$sel:permissions:CreateDataSet', createDataSet_permissions - A list of resource permissions on the dataset.

$sel:columnLevelPermissionRules:CreateDataSet', createDataSet_columnLevelPermissionRules - A set of one or more definitions of a ColumnLevelPermissionRule .

$sel:tags:CreateDataSet', createDataSet_tags - Contains a map of the key-value pairs for the resource tag or tags assigned to the dataset.

$sel:logicalTableMap:CreateDataSet', createDataSet_logicalTableMap - Configures the combination and transformation of the data from the physical tables.

$sel:awsAccountId:CreateDataSet', createDataSet_awsAccountId - The Amazon Web Services account ID.

$sel:dataSetId:CreateDataSet', createDataSet_dataSetId - An ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

$sel:name:CreateDataSet', createDataSet_name - The display name for the dataset.

$sel:physicalTableMap:CreateDataSet', createDataSet_physicalTableMap - Declares the physical tables that are available in the underlying data sources.

$sel:importMode:CreateDataSet', createDataSet_importMode - Indicates whether you want to import the data into SPICE.

Request Lenses

createDataSet_fieldFolders :: Lens' CreateDataSet (Maybe (HashMap Text FieldFolder))

The folder that contains fields and nested subfolders for your dataset.

createDataSet_columnGroups :: Lens' CreateDataSet (Maybe (NonEmpty ColumnGroup))

Groupings of columns that work together in certain Amazon QuickSight features. Currently, only geospatial hierarchy is supported.

createDataSet_dataSetUsageConfiguration :: Lens' CreateDataSet (Maybe DataSetUsageConfiguration)

Undocumented member.

createDataSet_rowLevelPermissionTagConfiguration :: Lens' CreateDataSet (Maybe RowLevelPermissionTagConfiguration)

The configuration of tags on a dataset to set row-level security. Row-level security tags are currently supported for anonymous embedding only.

createDataSet_rowLevelPermissionDataSet :: Lens' CreateDataSet (Maybe RowLevelPermissionDataSet)

The row-level security configuration for the data that you want to create.

createDataSet_permissions :: Lens' CreateDataSet (Maybe (NonEmpty ResourcePermission))

A list of resource permissions on the dataset.

createDataSet_columnLevelPermissionRules :: Lens' CreateDataSet (Maybe (NonEmpty ColumnLevelPermissionRule))

A set of one or more definitions of a ColumnLevelPermissionRule .

createDataSet_tags :: Lens' CreateDataSet (Maybe (NonEmpty Tag))

Contains a map of the key-value pairs for the resource tag or tags assigned to the dataset.

createDataSet_logicalTableMap :: Lens' CreateDataSet (Maybe (HashMap Text LogicalTable))

Configures the combination and transformation of the data from the physical tables.

createDataSet_awsAccountId :: Lens' CreateDataSet Text

The Amazon Web Services account ID.

createDataSet_dataSetId :: Lens' CreateDataSet Text

An ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

createDataSet_name :: Lens' CreateDataSet Text

The display name for the dataset.

createDataSet_physicalTableMap :: Lens' CreateDataSet (HashMap Text PhysicalTable)

Declares the physical tables that are available in the underlying data sources.

createDataSet_importMode :: Lens' CreateDataSet DataSetImportMode

Indicates whether you want to import the data into SPICE.

Destructuring the Response

data CreateDataSetResponse

See: newCreateDataSetResponse smart constructor.

Constructors

CreateDataSetResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

ingestionArn :: Maybe Text

The ARN for the ingestion, which is triggered as a result of dataset creation if the import mode is SPICE.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset.

ingestionId :: Maybe Text

The ID of the ingestion, which is triggered as a result of dataset creation if the import mode is SPICE.

dataSetId :: Maybe Text

The ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

status :: Int

The HTTP status of the request.

newCreateDataSetResponse

Arguments

:: Int
$sel:status:CreateDataSetResponse'

-> CreateDataSetResponse

Create a value of CreateDataSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateDataSetResponse', createDataSetResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:ingestionArn:CreateDataSetResponse', createDataSetResponse_ingestionArn - The ARN for the ingestion, which is triggered as a result of dataset creation if the import mode is SPICE.

$sel:arn:CreateDataSetResponse', createDataSetResponse_arn - The Amazon Resource Name (ARN) of the dataset.

$sel:ingestionId:CreateDataSetResponse', createDataSetResponse_ingestionId - The ID of the ingestion, which is triggered as a result of dataset creation if the import mode is SPICE.

$sel:dataSetId:CreateDataSet', createDataSetResponse_dataSetId - The ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

$sel:status:CreateDataSetResponse', createDataSetResponse_status - The HTTP status of the request.

Response Lenses

createDataSetResponse_requestId :: Lens' CreateDataSetResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

createDataSetResponse_ingestionArn :: Lens' CreateDataSetResponse (Maybe Text)

The ARN for the ingestion, which is triggered as a result of dataset creation if the import mode is SPICE.

createDataSetResponse_arn :: Lens' CreateDataSetResponse (Maybe Text)

The Amazon Resource Name (ARN) of the dataset.

createDataSetResponse_ingestionId :: Lens' CreateDataSetResponse (Maybe Text)

The ID of the ingestion, which is triggered as a result of dataset creation if the import mode is SPICE.

createDataSetResponse_dataSetId :: Lens' CreateDataSetResponse (Maybe Text)

The ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

createDataSetResponse_status :: Lens' CreateDataSetResponse Int

The HTTP status of the request.