Amazonka.QuickSight.CreateFolder

Description

Creates an empty shared folder.

Creating a Request

data CreateFolder

See: newCreateFolder smart constructor.

Constructors

CreateFolder'

Fields

parentFolderArn :: Maybe Text

The Amazon Resource Name (ARN) for the parent folder.

ParentFolderArn can be null. An empty parentFolderArn creates a root-level folder.

name :: Maybe Text

The name of the folder.

permissions :: Maybe (NonEmpty ResourcePermission)

A structure that describes the principals and the resource-level permissions of a folder.

To specify no permissions, omit Permissions.

folderType :: Maybe FolderType

The type of folder. By default, folderType is SHARED.

tags :: Maybe (NonEmpty Tag)

Tags for the folder.

awsAccountId :: Text

The AWS Account ID.

folderId :: Text

The folder ID.

newCreateFolder

Arguments

:: Text
$sel:awsAccountId:CreateFolder'

-> Text
$sel:folderId:CreateFolder'

-> CreateFolder

Create a value of CreateFolder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentFolderArn:CreateFolder', createFolder_parentFolderArn - The Amazon Resource Name (ARN) for the parent folder.

ParentFolderArn can be null. An empty parentFolderArn creates a root-level folder.

$sel:name:CreateFolder', createFolder_name - The name of the folder.

$sel:permissions:CreateFolder', createFolder_permissions - A structure that describes the principals and the resource-level permissions of a folder.

To specify no permissions, omit Permissions.

$sel:folderType:CreateFolder', createFolder_folderType - The type of folder. By default, folderType is SHARED.

$sel:tags:CreateFolder', createFolder_tags - Tags for the folder.

$sel:awsAccountId:CreateFolder', createFolder_awsAccountId - The AWS Account ID.

$sel:folderId:CreateFolder', createFolder_folderId - The folder ID.

Request Lenses

createFolder_parentFolderArn :: Lens' CreateFolder (Maybe Text)

The Amazon Resource Name (ARN) for the parent folder.

ParentFolderArn can be null. An empty parentFolderArn creates a root-level folder.

createFolder_name :: Lens' CreateFolder (Maybe Text)

The name of the folder.

createFolder_permissions :: Lens' CreateFolder (Maybe (NonEmpty ResourcePermission))

A structure that describes the principals and the resource-level permissions of a folder.

To specify no permissions, omit Permissions.

createFolder_folderType :: Lens' CreateFolder (Maybe FolderType)

The type of folder. By default, folderType is SHARED.

createFolder_tags :: Lens' CreateFolder (Maybe (NonEmpty Tag))

Tags for the folder.

createFolder_awsAccountId :: Lens' CreateFolder Text

The AWS Account ID.

createFolder_folderId :: Lens' CreateFolder Text

The folder ID.

Destructuring the Response

data CreateFolderResponse

See: newCreateFolderResponse smart constructor.

Constructors

CreateFolderResponse'

Fields

requestId :: Maybe Text

The request ID for the newly created folder.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the newly created folder.

folderId :: Maybe Text

The folder ID for the newly created folder.

status :: Int

The status of the newly created folder. If succeeded, the status is SC_OK (200).

newCreateFolderResponse

Arguments

:: Int
$sel:status:CreateFolderResponse'

-> CreateFolderResponse

Create a value of CreateFolderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateFolderResponse', createFolderResponse_requestId - The request ID for the newly created folder.

$sel:arn:CreateFolderResponse', createFolderResponse_arn - The Amazon Resource Name (ARN) for the newly created folder.

$sel:folderId:CreateFolder', createFolderResponse_folderId - The folder ID for the newly created folder.

$sel:status:CreateFolderResponse', createFolderResponse_status - The status of the newly created folder. If succeeded, the status is SC_OK (200).

Response Lenses

createFolderResponse_requestId :: Lens' CreateFolderResponse (Maybe Text)

The request ID for the newly created folder.

createFolderResponse_arn :: Lens' CreateFolderResponse (Maybe Text)

The Amazon Resource Name (ARN) for the newly created folder.

createFolderResponse_folderId :: Lens' CreateFolderResponse (Maybe Text)

The folder ID for the newly created folder.

createFolderResponse_status :: Lens' CreateFolderResponse Int

The status of the newly created folder. If succeeded, the status is SC_OK (200).