Amazonka.QuickSight.CreateFolderMembership

Description

Adds an asset, such as a dashboard, analysis, or dataset into a folder.

Creating a Request

data CreateFolderMembership

See: newCreateFolderMembership smart constructor.

Constructors

CreateFolderMembership'

Fields

awsAccountId :: Text

The AWS Account ID.

folderId :: Text

The folder ID.

memberId :: Text

The ID of the asset (the dashboard, analysis, or dataset).

memberType :: MemberType

The type of the member, including DASHBOARD, ANALYSIS, and DATASET.

newCreateFolderMembership

Arguments

:: Text
$sel:awsAccountId:CreateFolderMembership'

-> Text
$sel:folderId:CreateFolderMembership'

-> Text
$sel:memberId:CreateFolderMembership'

-> MemberType
$sel:memberType:CreateFolderMembership'

-> CreateFolderMembership

Create a value of CreateFolderMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:CreateFolderMembership', createFolderMembership_awsAccountId - The AWS Account ID.

$sel:folderId:CreateFolderMembership', createFolderMembership_folderId - The folder ID.

$sel:memberId:CreateFolderMembership', createFolderMembership_memberId - The ID of the asset (the dashboard, analysis, or dataset).

$sel:memberType:CreateFolderMembership', createFolderMembership_memberType - The type of the member, including DASHBOARD, ANALYSIS, and DATASET.

Request Lenses

createFolderMembership_awsAccountId :: Lens' CreateFolderMembership Text

The AWS Account ID.

createFolderMembership_folderId :: Lens' CreateFolderMembership Text

The folder ID.

createFolderMembership_memberId :: Lens' CreateFolderMembership Text

The ID of the asset (the dashboard, analysis, or dataset).

createFolderMembership_memberType :: Lens' CreateFolderMembership MemberType

The type of the member, including DASHBOARD, ANALYSIS, and DATASET.

Destructuring the Response

data CreateFolderMembershipResponse

See: newCreateFolderMembershipResponse smart constructor.

Constructors

CreateFolderMembershipResponse'

Fields

requestId :: Maybe Text

The request ID.

status :: Maybe Int

The status of the folder membership. If succeeded, the status is SC_OK (200).

folderMember :: Maybe FolderMember

Information about the member in the folder.

httpStatus :: Int

The response's http status code.

newCreateFolderMembershipResponse

Arguments

:: Int
$sel:httpStatus:CreateFolderMembershipResponse'

-> CreateFolderMembershipResponse

Create a value of CreateFolderMembershipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateFolderMembershipResponse', createFolderMembershipResponse_requestId - The request ID.

$sel:status:CreateFolderMembershipResponse', createFolderMembershipResponse_status - The status of the folder membership. If succeeded, the status is SC_OK (200).

$sel:folderMember:CreateFolderMembershipResponse', createFolderMembershipResponse_folderMember - Information about the member in the folder.

$sel:httpStatus:CreateFolderMembershipResponse', createFolderMembershipResponse_httpStatus - The response's http status code.

Response Lenses

createFolderMembershipResponse_requestId :: Lens' CreateFolderMembershipResponse (Maybe Text)

The request ID.

createFolderMembershipResponse_status :: Lens' CreateFolderMembershipResponse (Maybe Int)

The status of the folder membership. If succeeded, the status is SC_OK (200).

createFolderMembershipResponse_folderMember :: Lens' CreateFolderMembershipResponse (Maybe FolderMember)

Information about the member in the folder.

createFolderMembershipResponse_httpStatus :: Lens' CreateFolderMembershipResponse Int

The response's http status code.