Amazonka.QuickSight.CreateIAMPolicyAssignment

Description

Creates an assignment with one specified IAMpolicy, identified by its Amazon Resource Name (ARN). This policy assignment is attached to the specified groups or users of Amazon QuickSight. Assignment names are unique per Amazon Web Services account. To avoid overwriting rules in other namespaces, use assignment names that are unique.

Creating a Request

data CreateIAMPolicyAssignment

See: newCreateIAMPolicyAssignment smart constructor.

Constructors

CreateIAMPolicyAssignment'

Fields

policyArn :: Maybe Text

The ARN for the IAMpolicy to apply to the Amazon QuickSight users and groups specified in this assignment.

identities :: Maybe (HashMap Text [Text])

The Amazon QuickSight users, groups, or both that you want to assign the policy to.

awsAccountId :: Text

The ID of the Amazon Web Services account where you want to assign an IAMpolicy to Amazon QuickSight users or groups.

assignmentName :: Text

The name of the assignment, also called a rule. It must be unique within an Amazon Web Services account.

assignmentStatus :: AssignmentStatus

The status of the assignment. Possible values are as follows:

ENABLED - Anything specified in this assignment is used when creating the data source.

DISABLED - This assignment isn't used when creating the data source.

DRAFT - This assignment is an unfinished draft and isn't used when creating the data source.

namespace :: Text

The namespace that contains the assignment.

newCreateIAMPolicyAssignment

Arguments

:: Text
$sel:awsAccountId:CreateIAMPolicyAssignment'

-> Text
$sel:assignmentName:CreateIAMPolicyAssignment'

-> AssignmentStatus
$sel:assignmentStatus:CreateIAMPolicyAssignment'

-> Text
$sel:namespace:CreateIAMPolicyAssignment'

-> CreateIAMPolicyAssignment

Create a value of CreateIAMPolicyAssignment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyArn:CreateIAMPolicyAssignment', createIAMPolicyAssignment_policyArn - The ARN for the IAMpolicy to apply to the Amazon QuickSight users and groups specified in this assignment.

$sel:identities:CreateIAMPolicyAssignment', createIAMPolicyAssignment_identities - The Amazon QuickSight users, groups, or both that you want to assign the policy to.

$sel:awsAccountId:CreateIAMPolicyAssignment', createIAMPolicyAssignment_awsAccountId - The ID of the Amazon Web Services account where you want to assign an IAMpolicy to Amazon QuickSight users or groups.

$sel:assignmentName:CreateIAMPolicyAssignment', createIAMPolicyAssignment_assignmentName - The name of the assignment, also called a rule. It must be unique within an Amazon Web Services account.

$sel:assignmentStatus:CreateIAMPolicyAssignment', createIAMPolicyAssignment_assignmentStatus - The status of the assignment. Possible values are as follows:

ENABLED - Anything specified in this assignment is used when creating the data source.

DISABLED - This assignment isn't used when creating the data source.

DRAFT - This assignment is an unfinished draft and isn't used when creating the data source.

$sel:namespace:CreateIAMPolicyAssignment', createIAMPolicyAssignment_namespace - The namespace that contains the assignment.

Request Lenses

createIAMPolicyAssignment_policyArn :: Lens' CreateIAMPolicyAssignment (Maybe Text)

The ARN for the IAMpolicy to apply to the Amazon QuickSight users and groups specified in this assignment.

createIAMPolicyAssignment_identities :: Lens' CreateIAMPolicyAssignment (Maybe (HashMap Text [Text]))

The Amazon QuickSight users, groups, or both that you want to assign the policy to.

createIAMPolicyAssignment_awsAccountId :: Lens' CreateIAMPolicyAssignment Text

The ID of the Amazon Web Services account where you want to assign an IAMpolicy to Amazon QuickSight users or groups.

createIAMPolicyAssignment_assignmentName :: Lens' CreateIAMPolicyAssignment Text

The name of the assignment, also called a rule. It must be unique within an Amazon Web Services account.

createIAMPolicyAssignment_assignmentStatus :: Lens' CreateIAMPolicyAssignment AssignmentStatus

The status of the assignment. Possible values are as follows:

ENABLED - Anything specified in this assignment is used when creating the data source.

DISABLED - This assignment isn't used when creating the data source.

DRAFT - This assignment is an unfinished draft and isn't used when creating the data source.

createIAMPolicyAssignment_namespace :: Lens' CreateIAMPolicyAssignment Text

The namespace that contains the assignment.

Destructuring the Response

data CreateIAMPolicyAssignmentResponse

See: newCreateIAMPolicyAssignmentResponse smart constructor.

Constructors

CreateIAMPolicyAssignmentResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

assignmentName :: Maybe Text

The name of the assignment. This name must be unique within the Amazon Web Services account.

assignmentStatus :: Maybe AssignmentStatus

The status of the assignment. Possible values are as follows:

ENABLED - Anything specified in this assignment is used when creating the data source.

DISABLED - This assignment isn't used when creating the data source.

DRAFT - This assignment is an unfinished draft and isn't used when creating the data source.

policyArn :: Maybe Text

The ARN for the IAMpolicy that is applied to the Amazon QuickSight users and groups specified in this assignment.

identities :: Maybe (HashMap Text [Text])

The Amazon QuickSight users, groups, or both that the IAMpolicy is assigned to.

assignmentId :: Maybe Text

The ID for the assignment.

status :: Int

The HTTP status of the request.

newCreateIAMPolicyAssignmentResponse

Arguments

:: Int
$sel:status:CreateIAMPolicyAssignmentResponse'

-> CreateIAMPolicyAssignmentResponse

Create a value of CreateIAMPolicyAssignmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateIAMPolicyAssignmentResponse', createIAMPolicyAssignmentResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:assignmentName:CreateIAMPolicyAssignment', createIAMPolicyAssignmentResponse_assignmentName - The name of the assignment. This name must be unique within the Amazon Web Services account.

$sel:assignmentStatus:CreateIAMPolicyAssignment', createIAMPolicyAssignmentResponse_assignmentStatus - The status of the assignment. Possible values are as follows:

ENABLED - Anything specified in this assignment is used when creating the data source.

DISABLED - This assignment isn't used when creating the data source.

DRAFT - This assignment is an unfinished draft and isn't used when creating the data source.

$sel:policyArn:CreateIAMPolicyAssignment', createIAMPolicyAssignmentResponse_policyArn - The ARN for the IAMpolicy that is applied to the Amazon QuickSight users and groups specified in this assignment.

$sel:identities:CreateIAMPolicyAssignment', createIAMPolicyAssignmentResponse_identities - The Amazon QuickSight users, groups, or both that the IAMpolicy is assigned to.

$sel:assignmentId:CreateIAMPolicyAssignmentResponse', createIAMPolicyAssignmentResponse_assignmentId - The ID for the assignment.

$sel:status:CreateIAMPolicyAssignmentResponse', createIAMPolicyAssignmentResponse_status - The HTTP status of the request.

Response Lenses

createIAMPolicyAssignmentResponse_requestId :: Lens' CreateIAMPolicyAssignmentResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

createIAMPolicyAssignmentResponse_assignmentName :: Lens' CreateIAMPolicyAssignmentResponse (Maybe Text)

The name of the assignment. This name must be unique within the Amazon Web Services account.

createIAMPolicyAssignmentResponse_assignmentStatus :: Lens' CreateIAMPolicyAssignmentResponse (Maybe AssignmentStatus)

The status of the assignment. Possible values are as follows:

ENABLED - Anything specified in this assignment is used when creating the data source.

DISABLED - This assignment isn't used when creating the data source.

DRAFT - This assignment is an unfinished draft and isn't used when creating the data source.

createIAMPolicyAssignmentResponse_policyArn :: Lens' CreateIAMPolicyAssignmentResponse (Maybe Text)

The ARN for the IAMpolicy that is applied to the Amazon QuickSight users and groups specified in this assignment.

createIAMPolicyAssignmentResponse_identities :: Lens' CreateIAMPolicyAssignmentResponse (Maybe (HashMap Text [Text]))

The Amazon QuickSight users, groups, or both that the IAMpolicy is assigned to.

createIAMPolicyAssignmentResponse_assignmentId :: Lens' CreateIAMPolicyAssignmentResponse (Maybe Text)

The ID for the assignment.

createIAMPolicyAssignmentResponse_status :: Lens' CreateIAMPolicyAssignmentResponse Int

The HTTP status of the request.