Amazonka.QuickSight.CreateTemplate

Description

Creates a template from an existing Amazon QuickSight analysis or template. You can use the resulting template to create a dashboard.

A template is an entity in Amazon QuickSight that encapsulates the metadata required to create an analysis and that you can use to create s dashboard. A template adds a layer of abstraction by using placeholders to replace the dataset associated with the analysis. You can use templates to create dashboards by replacing dataset placeholders with datasets that follow the same schema that was used to create the source analysis and template.

Creating a Request

data CreateTemplate

See: newCreateTemplate smart constructor.

Constructors

CreateTemplate'

Fields

versionDescription :: Maybe Text

A description of the current template version being created. This API operation creates the first version of the template. Every time UpdateTemplate is called, a new version is created. Each version of the template maintains a description of the version in the VersionDescription field.

name :: Maybe Text

A display name for the template.

permissions :: Maybe (NonEmpty ResourcePermission)

A list of resource permissions to be set on the template.

tags :: Maybe (NonEmpty Tag)

Contains a map of the key-value pairs for the resource tag or tags assigned to the resource.

awsAccountId :: Text

The ID for the Amazon Web Services account that the group is in. You use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

templateId :: Text

An ID for the template that you want to create. This template is unique per Amazon Web Services Region; in each Amazon Web Services account.

sourceEntity :: TemplateSourceEntity

The entity that you are using as a source when you create the template. In SourceEntity, you specify the type of object you're using as source: SourceTemplate for a template or SourceAnalysis for an analysis. Both of these require an Amazon Resource Name (ARN). For SourceTemplate, specify the ARN of the source template. For SourceAnalysis, specify the ARN of the source analysis. The SourceTemplate ARN can contain any Amazon Web Services account and any Amazon QuickSight-supported Amazon Web Services Region.

Use the DataSetReferences entity within SourceTemplate or SourceAnalysis to list the replacement datasets for the placeholders listed in the original. The schema in each dataset must match its placeholder.

newCreateTemplate

Arguments

:: Text
$sel:awsAccountId:CreateTemplate'

-> Text
$sel:templateId:CreateTemplate'

-> TemplateSourceEntity
$sel:sourceEntity:CreateTemplate'

-> CreateTemplate

Create a value of CreateTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionDescription:CreateTemplate', createTemplate_versionDescription - A description of the current template version being created. This API operation creates the first version of the template. Every time UpdateTemplate is called, a new version is created. Each version of the template maintains a description of the version in the VersionDescription field.

$sel:name:CreateTemplate', createTemplate_name - A display name for the template.

$sel:permissions:CreateTemplate', createTemplate_permissions - A list of resource permissions to be set on the template.

$sel:tags:CreateTemplate', createTemplate_tags - Contains a map of the key-value pairs for the resource tag or tags assigned to the resource.

$sel:awsAccountId:CreateTemplate', createTemplate_awsAccountId - The ID for the Amazon Web Services account that the group is in. You use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

$sel:templateId:CreateTemplate', createTemplate_templateId - An ID for the template that you want to create. This template is unique per Amazon Web Services Region; in each Amazon Web Services account.

$sel:sourceEntity:CreateTemplate', createTemplate_sourceEntity - The entity that you are using as a source when you create the template. In SourceEntity, you specify the type of object you're using as source: SourceTemplate for a template or SourceAnalysis for an analysis. Both of these require an Amazon Resource Name (ARN). For SourceTemplate, specify the ARN of the source template. For SourceAnalysis, specify the ARN of the source analysis. The SourceTemplate ARN can contain any Amazon Web Services account and any Amazon QuickSight-supported Amazon Web Services Region.

Use the DataSetReferences entity within SourceTemplate or SourceAnalysis to list the replacement datasets for the placeholders listed in the original. The schema in each dataset must match its placeholder.

Request Lenses

createTemplate_versionDescription :: Lens' CreateTemplate (Maybe Text)

A description of the current template version being created. This API operation creates the first version of the template. Every time UpdateTemplate is called, a new version is created. Each version of the template maintains a description of the version in the VersionDescription field.

createTemplate_name :: Lens' CreateTemplate (Maybe Text)

A display name for the template.

createTemplate_permissions :: Lens' CreateTemplate (Maybe (NonEmpty ResourcePermission))

A list of resource permissions to be set on the template.

createTemplate_tags :: Lens' CreateTemplate (Maybe (NonEmpty Tag))

Contains a map of the key-value pairs for the resource tag or tags assigned to the resource.

createTemplate_awsAccountId :: Lens' CreateTemplate Text

The ID for the Amazon Web Services account that the group is in. You use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

createTemplate_templateId :: Lens' CreateTemplate Text

An ID for the template that you want to create. This template is unique per Amazon Web Services Region; in each Amazon Web Services account.

createTemplate_sourceEntity :: Lens' CreateTemplate TemplateSourceEntity

The entity that you are using as a source when you create the template. In SourceEntity, you specify the type of object you're using as source: SourceTemplate for a template or SourceAnalysis for an analysis. Both of these require an Amazon Resource Name (ARN). For SourceTemplate, specify the ARN of the source template. For SourceAnalysis, specify the ARN of the source analysis. The SourceTemplate ARN can contain any Amazon Web Services account and any Amazon QuickSight-supported Amazon Web Services Region.

Use the DataSetReferences entity within SourceTemplate or SourceAnalysis to list the replacement datasets for the placeholders listed in the original. The schema in each dataset must match its placeholder.

Destructuring the Response

data CreateTemplateResponse

See: newCreateTemplateResponse smart constructor.

Constructors

CreateTemplateResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

arn :: Maybe Text

The ARN for the template.

creationStatus :: Maybe ResourceStatus

The template creation status.

templateId :: Maybe Text

The ID of the template.

versionArn :: Maybe Text

The ARN for the template, including the version information of the first version.

status :: Int

The HTTP status of the request.

newCreateTemplateResponse

Arguments

:: Int
$sel:status:CreateTemplateResponse'

-> CreateTemplateResponse

Create a value of CreateTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateTemplateResponse', createTemplateResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:arn:CreateTemplateResponse', createTemplateResponse_arn - The ARN for the template.

$sel:creationStatus:CreateTemplateResponse', createTemplateResponse_creationStatus - The template creation status.

$sel:templateId:CreateTemplate', createTemplateResponse_templateId - The ID of the template.

$sel:versionArn:CreateTemplateResponse', createTemplateResponse_versionArn - The ARN for the template, including the version information of the first version.

$sel:status:CreateTemplateResponse', createTemplateResponse_status - The HTTP status of the request.

Response Lenses

createTemplateResponse_requestId :: Lens' CreateTemplateResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

createTemplateResponse_arn :: Lens' CreateTemplateResponse (Maybe Text)

The ARN for the template.

createTemplateResponse_creationStatus :: Lens' CreateTemplateResponse (Maybe ResourceStatus)

The template creation status.

createTemplateResponse_templateId :: Lens' CreateTemplateResponse (Maybe Text)

The ID of the template.

createTemplateResponse_versionArn :: Lens' CreateTemplateResponse (Maybe Text)

The ARN for the template, including the version information of the first version.

createTemplateResponse_status :: Lens' CreateTemplateResponse Int

The HTTP status of the request.