Amazonka.QuickSight.DeleteGroupMembership

Description

Removes a user from a group so that the user is no longer a member of the group.

Creating a Request

data DeleteGroupMembership

See: newDeleteGroupMembership smart constructor.

Constructors

DeleteGroupMembership'

Fields

memberName :: Text

The name of the user that you want to delete from the group membership.

groupName :: Text

The name of the group that you want to delete the user from.

awsAccountId :: Text

The ID for the Amazon Web Services account that the group is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

namespace :: Text

The namespace. Currently, you should set this to default.

newDeleteGroupMembership

Arguments

:: Text
$sel:memberName:DeleteGroupMembership'

-> Text
$sel:groupName:DeleteGroupMembership'

-> Text
$sel:awsAccountId:DeleteGroupMembership'

-> Text
$sel:namespace:DeleteGroupMembership'

-> DeleteGroupMembership

Create a value of DeleteGroupMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberName:DeleteGroupMembership', deleteGroupMembership_memberName - The name of the user that you want to delete from the group membership.

$sel:groupName:DeleteGroupMembership', deleteGroupMembership_groupName - The name of the group that you want to delete the user from.

$sel:awsAccountId:DeleteGroupMembership', deleteGroupMembership_awsAccountId - The ID for the Amazon Web Services account that the group is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

$sel:namespace:DeleteGroupMembership', deleteGroupMembership_namespace - The namespace. Currently, you should set this to default.

Request Lenses

deleteGroupMembership_memberName :: Lens' DeleteGroupMembership Text

The name of the user that you want to delete from the group membership.

deleteGroupMembership_groupName :: Lens' DeleteGroupMembership Text

The name of the group that you want to delete the user from.

deleteGroupMembership_awsAccountId :: Lens' DeleteGroupMembership Text

The ID for the Amazon Web Services account that the group is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

deleteGroupMembership_namespace :: Lens' DeleteGroupMembership Text

The namespace. Currently, you should set this to default.

Destructuring the Response

data DeleteGroupMembershipResponse

See: newDeleteGroupMembershipResponse smart constructor.

Constructors

DeleteGroupMembershipResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

status :: Int

The HTTP status of the request.

newDeleteGroupMembershipResponse

Arguments

:: Int
$sel:status:DeleteGroupMembershipResponse'

-> DeleteGroupMembershipResponse

Create a value of DeleteGroupMembershipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DeleteGroupMembershipResponse', deleteGroupMembershipResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:status:DeleteGroupMembershipResponse', deleteGroupMembershipResponse_status - The HTTP status of the request.

Response Lenses

deleteGroupMembershipResponse_requestId :: Lens' DeleteGroupMembershipResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

deleteGroupMembershipResponse_status :: Lens' DeleteGroupMembershipResponse Int

The HTTP status of the request.