Amazonka.QuickSight.DeleteThemeAlias

Description

Deletes the version of the theme that the specified theme alias points to. If you provide a specific alias, you delete the version of the theme that the alias points to.

Creating a Request

data DeleteThemeAlias

See: newDeleteThemeAlias smart constructor.

Constructors

DeleteThemeAlias'

Fields

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the theme alias to delete.

themeId :: Text

The ID for the theme that the specified alias is for.

aliasName :: Text

The unique name for the theme alias to delete.

newDeleteThemeAlias

Arguments

:: Text
$sel:awsAccountId:DeleteThemeAlias'

-> Text
$sel:themeId:DeleteThemeAlias'

-> Text
$sel:aliasName:DeleteThemeAlias'

-> DeleteThemeAlias

Create a value of DeleteThemeAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:DeleteThemeAlias', deleteThemeAlias_awsAccountId - The ID of the Amazon Web Services account that contains the theme alias to delete.

$sel:themeId:DeleteThemeAlias', deleteThemeAlias_themeId - The ID for the theme that the specified alias is for.

$sel:aliasName:DeleteThemeAlias', deleteThemeAlias_aliasName - The unique name for the theme alias to delete.

Request Lenses

deleteThemeAlias_awsAccountId :: Lens' DeleteThemeAlias Text

The ID of the Amazon Web Services account that contains the theme alias to delete.

deleteThemeAlias_themeId :: Lens' DeleteThemeAlias Text

The ID for the theme that the specified alias is for.

deleteThemeAlias_aliasName :: Lens' DeleteThemeAlias Text

The unique name for the theme alias to delete.

Destructuring the Response

data DeleteThemeAliasResponse

See: newDeleteThemeAliasResponse smart constructor.

Constructors

DeleteThemeAliasResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

themeId :: Maybe Text

An ID for the theme associated with the deletion.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the theme resource using the deleted alias.

aliasName :: Maybe Text

The name for the theme alias.

status :: Int

The HTTP status of the request.

newDeleteThemeAliasResponse

Arguments

:: Int
$sel:status:DeleteThemeAliasResponse'

-> DeleteThemeAliasResponse

Create a value of DeleteThemeAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DeleteThemeAliasResponse', deleteThemeAliasResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:themeId:DeleteThemeAlias', deleteThemeAliasResponse_themeId - An ID for the theme associated with the deletion.

$sel:arn:DeleteThemeAliasResponse', deleteThemeAliasResponse_arn - The Amazon Resource Name (ARN) of the theme resource using the deleted alias.

$sel:aliasName:DeleteThemeAlias', deleteThemeAliasResponse_aliasName - The name for the theme alias.

$sel:status:DeleteThemeAliasResponse', deleteThemeAliasResponse_status - The HTTP status of the request.

Response Lenses

deleteThemeAliasResponse_requestId :: Lens' DeleteThemeAliasResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

deleteThemeAliasResponse_themeId :: Lens' DeleteThemeAliasResponse (Maybe Text)

An ID for the theme associated with the deletion.

deleteThemeAliasResponse_arn :: Lens' DeleteThemeAliasResponse (Maybe Text)

The Amazon Resource Name (ARN) of the theme resource using the deleted alias.

deleteThemeAliasResponse_aliasName :: Lens' DeleteThemeAliasResponse (Maybe Text)

The name for the theme alias.

deleteThemeAliasResponse_status :: Lens' DeleteThemeAliasResponse Int

The HTTP status of the request.