Amazonka.QuickSight.DeleteUserByPrincipalId

Description

Deletes a user identified by its principal ID.

Creating a Request

data DeleteUserByPrincipalId

See: newDeleteUserByPrincipalId smart constructor.

Constructors

DeleteUserByPrincipalId'

Fields

principalId :: Text

The principal ID of the user.

awsAccountId :: Text

The ID for the Amazon Web Services account that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

namespace :: Text

The namespace. Currently, you should set this to default.

newDeleteUserByPrincipalId

Arguments

:: Text
$sel:principalId:DeleteUserByPrincipalId'

-> Text
$sel:awsAccountId:DeleteUserByPrincipalId'

-> Text
$sel:namespace:DeleteUserByPrincipalId'

-> DeleteUserByPrincipalId

Create a value of DeleteUserByPrincipalId with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principalId:DeleteUserByPrincipalId', deleteUserByPrincipalId_principalId - The principal ID of the user.

$sel:awsAccountId:DeleteUserByPrincipalId', deleteUserByPrincipalId_awsAccountId - The ID for the Amazon Web Services account that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

$sel:namespace:DeleteUserByPrincipalId', deleteUserByPrincipalId_namespace - The namespace. Currently, you should set this to default.

Request Lenses

deleteUserByPrincipalId_principalId :: Lens' DeleteUserByPrincipalId Text

The principal ID of the user.

deleteUserByPrincipalId_awsAccountId :: Lens' DeleteUserByPrincipalId Text

The ID for the Amazon Web Services account that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

deleteUserByPrincipalId_namespace :: Lens' DeleteUserByPrincipalId Text

The namespace. Currently, you should set this to default.

Destructuring the Response

data DeleteUserByPrincipalIdResponse

See: newDeleteUserByPrincipalIdResponse smart constructor.

Constructors

DeleteUserByPrincipalIdResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

status :: Int

The HTTP status of the request.

newDeleteUserByPrincipalIdResponse

Arguments

:: Int
$sel:status:DeleteUserByPrincipalIdResponse'

-> DeleteUserByPrincipalIdResponse

Create a value of DeleteUserByPrincipalIdResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DeleteUserByPrincipalIdResponse', deleteUserByPrincipalIdResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:status:DeleteUserByPrincipalIdResponse', deleteUserByPrincipalIdResponse_status - The HTTP status of the request.

Response Lenses

deleteUserByPrincipalIdResponse_requestId :: Lens' DeleteUserByPrincipalIdResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

deleteUserByPrincipalIdResponse_status :: Lens' DeleteUserByPrincipalIdResponse Int

The HTTP status of the request.