Amazonka.QuickSight.DescribeDashboardPermissions

Description

Describes read and write permissions for a dashboard.

Creating a Request

data DescribeDashboardPermissions

See: newDescribeDashboardPermissions smart constructor.

Constructors

DescribeDashboardPermissions'

Fields

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the dashboard that you're describing permissions for.

dashboardId :: Text

The ID for the dashboard, also added to the IAMpolicy.

newDescribeDashboardPermissions

Arguments

:: Text
$sel:awsAccountId:DescribeDashboardPermissions'

-> Text
$sel:dashboardId:DescribeDashboardPermissions'

-> DescribeDashboardPermissions

Create a value of DescribeDashboardPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:DescribeDashboardPermissions', describeDashboardPermissions_awsAccountId - The ID of the Amazon Web Services account that contains the dashboard that you're describing permissions for.

$sel:dashboardId:DescribeDashboardPermissions', describeDashboardPermissions_dashboardId - The ID for the dashboard, also added to the IAMpolicy.

Request Lenses

describeDashboardPermissions_awsAccountId :: Lens' DescribeDashboardPermissions Text

The ID of the Amazon Web Services account that contains the dashboard that you're describing permissions for.

describeDashboardPermissions_dashboardId :: Lens' DescribeDashboardPermissions Text

The ID for the dashboard, also added to the IAMpolicy.

Destructuring the Response

data DescribeDashboardPermissionsResponse

See: newDescribeDashboardPermissionsResponse smart constructor.

Constructors

DescribeDashboardPermissionsResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

dashboardId :: Maybe Text

The ID for the dashboard.

permissions :: Maybe (NonEmpty ResourcePermission)

A structure that contains the permissions for the dashboard.

dashboardArn :: Maybe Text

The Amazon Resource Name (ARN) of the dashboard.

status :: Int

The HTTP status of the request.

newDescribeDashboardPermissionsResponse

Arguments

:: Int
$sel:status:DescribeDashboardPermissionsResponse'

-> DescribeDashboardPermissionsResponse

Create a value of DescribeDashboardPermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DescribeDashboardPermissionsResponse', describeDashboardPermissionsResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:dashboardId:DescribeDashboardPermissions', describeDashboardPermissionsResponse_dashboardId - The ID for the dashboard.

$sel:permissions:DescribeDashboardPermissionsResponse', describeDashboardPermissionsResponse_permissions - A structure that contains the permissions for the dashboard.

$sel:dashboardArn:DescribeDashboardPermissionsResponse', describeDashboardPermissionsResponse_dashboardArn - The Amazon Resource Name (ARN) of the dashboard.

$sel:status:DescribeDashboardPermissionsResponse', describeDashboardPermissionsResponse_status - The HTTP status of the request.

Response Lenses

describeDashboardPermissionsResponse_requestId :: Lens' DescribeDashboardPermissionsResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

describeDashboardPermissionsResponse_dashboardId :: Lens' DescribeDashboardPermissionsResponse (Maybe Text)

The ID for the dashboard.

describeDashboardPermissionsResponse_permissions :: Lens' DescribeDashboardPermissionsResponse (Maybe (NonEmpty ResourcePermission))

A structure that contains the permissions for the dashboard.

describeDashboardPermissionsResponse_dashboardArn :: Lens' DescribeDashboardPermissionsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the dashboard.

describeDashboardPermissionsResponse_status :: Lens' DescribeDashboardPermissionsResponse Int

The HTTP status of the request.