Amazonka.QuickSight.DescribeDataSetPermissions

Description

Describes the permissions on a dataset.

The permissions resource is arn:aws:quicksight:region:aws-account-id:dataset/data-set-id.

Creating a Request

data DescribeDataSetPermissions

See: newDescribeDataSetPermissions smart constructor.

Constructors

DescribeDataSetPermissions'

Fields

awsAccountId :: Text

The Amazon Web Services account ID.

dataSetId :: Text

The ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

newDescribeDataSetPermissions

Arguments

:: Text
$sel:awsAccountId:DescribeDataSetPermissions'

-> Text
$sel:dataSetId:DescribeDataSetPermissions'

-> DescribeDataSetPermissions

Create a value of DescribeDataSetPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:DescribeDataSetPermissions', describeDataSetPermissions_awsAccountId - The Amazon Web Services account ID.

$sel:dataSetId:DescribeDataSetPermissions', describeDataSetPermissions_dataSetId - The ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

Request Lenses

describeDataSetPermissions_awsAccountId :: Lens' DescribeDataSetPermissions Text

The Amazon Web Services account ID.

describeDataSetPermissions_dataSetId :: Lens' DescribeDataSetPermissions Text

The ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

Destructuring the Response

data DescribeDataSetPermissionsResponse

See: newDescribeDataSetPermissionsResponse smart constructor.

Constructors

DescribeDataSetPermissionsResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

dataSetArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset.

dataSetId :: Maybe Text

The ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

permissions :: Maybe (NonEmpty ResourcePermission)

A list of resource permissions on the dataset.

status :: Int

The HTTP status of the request.

newDescribeDataSetPermissionsResponse

Arguments

:: Int
$sel:status:DescribeDataSetPermissionsResponse'

-> DescribeDataSetPermissionsResponse

Create a value of DescribeDataSetPermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DescribeDataSetPermissionsResponse', describeDataSetPermissionsResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:dataSetArn:DescribeDataSetPermissionsResponse', describeDataSetPermissionsResponse_dataSetArn - The Amazon Resource Name (ARN) of the dataset.

$sel:dataSetId:DescribeDataSetPermissions', describeDataSetPermissionsResponse_dataSetId - The ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

$sel:permissions:DescribeDataSetPermissionsResponse', describeDataSetPermissionsResponse_permissions - A list of resource permissions on the dataset.

$sel:status:DescribeDataSetPermissionsResponse', describeDataSetPermissionsResponse_status - The HTTP status of the request.

Response Lenses

describeDataSetPermissionsResponse_requestId :: Lens' DescribeDataSetPermissionsResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

describeDataSetPermissionsResponse_dataSetArn :: Lens' DescribeDataSetPermissionsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the dataset.

describeDataSetPermissionsResponse_dataSetId :: Lens' DescribeDataSetPermissionsResponse (Maybe Text)

The ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

describeDataSetPermissionsResponse_permissions :: Lens' DescribeDataSetPermissionsResponse (Maybe (NonEmpty ResourcePermission))

A list of resource permissions on the dataset.

describeDataSetPermissionsResponse_status :: Lens' DescribeDataSetPermissionsResponse Int

The HTTP status of the request.