Amazonka.QuickSight.DescribeIpRestriction

Description

Provides a summary and status of IP Rules.

Creating a Request

data DescribeIpRestriction

See: newDescribeIpRestriction smart constructor.

Constructors

DescribeIpRestriction'

Fields

awsAccountId :: Text

Your AWS account ID.

newDescribeIpRestriction

Arguments

:: Text
$sel:awsAccountId:DescribeIpRestriction'

-> DescribeIpRestriction

Create a value of DescribeIpRestriction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:DescribeIpRestriction', describeIpRestriction_awsAccountId - Your AWS account ID.

Request Lenses

describeIpRestriction_awsAccountId :: Lens' DescribeIpRestriction Text

Your AWS account ID.

Destructuring the Response

data DescribeIpRestrictionResponse

See: newDescribeIpRestrictionResponse smart constructor.

Constructors

DescribeIpRestrictionResponse'

Fields

requestId :: Maybe Text

The ID of the describe request.

ipRestrictionRuleMap :: Maybe (HashMap Text Text)

Describes the IP rules with CIDR range and description.

enabled :: Maybe Bool

Whether or not IP rules are enabled.

awsAccountId :: Maybe Text

Your AWS account ID.

status :: Int

The status of a set of IP restrictions. A successful request returns a 200 status code.

newDescribeIpRestrictionResponse

Arguments

:: Int
$sel:status:DescribeIpRestrictionResponse'

-> DescribeIpRestrictionResponse

Create a value of DescribeIpRestrictionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DescribeIpRestrictionResponse', describeIpRestrictionResponse_requestId - The ID of the describe request.

$sel:ipRestrictionRuleMap:DescribeIpRestrictionResponse', describeIpRestrictionResponse_ipRestrictionRuleMap - Describes the IP rules with CIDR range and description.

$sel:enabled:DescribeIpRestrictionResponse', describeIpRestrictionResponse_enabled - Whether or not IP rules are enabled.

$sel:awsAccountId:DescribeIpRestriction', describeIpRestrictionResponse_awsAccountId - Your AWS account ID.

$sel:status:DescribeIpRestrictionResponse', describeIpRestrictionResponse_status - The status of a set of IP restrictions. A successful request returns a 200 status code.

Response Lenses

describeIpRestrictionResponse_requestId :: Lens' DescribeIpRestrictionResponse (Maybe Text)

The ID of the describe request.

describeIpRestrictionResponse_ipRestrictionRuleMap :: Lens' DescribeIpRestrictionResponse (Maybe (HashMap Text Text))

Describes the IP rules with CIDR range and description.

describeIpRestrictionResponse_enabled :: Lens' DescribeIpRestrictionResponse (Maybe Bool)

Whether or not IP rules are enabled.

describeIpRestrictionResponse_awsAccountId :: Lens' DescribeIpRestrictionResponse (Maybe Text)

Your AWS account ID.

describeIpRestrictionResponse_status :: Lens' DescribeIpRestrictionResponse Int

The status of a set of IP restrictions. A successful request returns a 200 status code.