Amazonka.QuickSight.DescribeTemplateAlias

Description

Describes the template alias for a template.

Creating a Request

data DescribeTemplateAlias

See: newDescribeTemplateAlias smart constructor.

Constructors

DescribeTemplateAlias'

Fields

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the template alias that you're describing.

templateId :: Text

The ID for the template.

aliasName :: Text

The name of the template alias that you want to describe. If you name a specific alias, you describe the version that the alias points to. You can specify the latest version of the template by providing the keyword $LATEST in the AliasName parameter. The keyword $PUBLISHED doesn't apply to templates.

newDescribeTemplateAlias

Arguments

:: Text
$sel:awsAccountId:DescribeTemplateAlias'

-> Text
$sel:templateId:DescribeTemplateAlias'

-> Text
$sel:aliasName:DescribeTemplateAlias'

-> DescribeTemplateAlias

Create a value of DescribeTemplateAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:DescribeTemplateAlias', describeTemplateAlias_awsAccountId - The ID of the Amazon Web Services account that contains the template alias that you're describing.

$sel:templateId:DescribeTemplateAlias', describeTemplateAlias_templateId - The ID for the template.

$sel:aliasName:DescribeTemplateAlias', describeTemplateAlias_aliasName - The name of the template alias that you want to describe. If you name a specific alias, you describe the version that the alias points to. You can specify the latest version of the template by providing the keyword $LATEST in the AliasName parameter. The keyword $PUBLISHED doesn't apply to templates.

Request Lenses

describeTemplateAlias_awsAccountId :: Lens' DescribeTemplateAlias Text

The ID of the Amazon Web Services account that contains the template alias that you're describing.

describeTemplateAlias_templateId :: Lens' DescribeTemplateAlias Text

The ID for the template.

describeTemplateAlias_aliasName :: Lens' DescribeTemplateAlias Text

The name of the template alias that you want to describe. If you name a specific alias, you describe the version that the alias points to. You can specify the latest version of the template by providing the keyword $LATEST in the AliasName parameter. The keyword $PUBLISHED doesn't apply to templates.

Destructuring the Response

data DescribeTemplateAliasResponse

See: newDescribeTemplateAliasResponse smart constructor.

Constructors

DescribeTemplateAliasResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

templateAlias :: Maybe TemplateAlias

Information about the template alias.

status :: Int

The HTTP status of the request.

newDescribeTemplateAliasResponse

Arguments

:: Int
$sel:status:DescribeTemplateAliasResponse'

-> DescribeTemplateAliasResponse

Create a value of DescribeTemplateAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DescribeTemplateAliasResponse', describeTemplateAliasResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:templateAlias:DescribeTemplateAliasResponse', describeTemplateAliasResponse_templateAlias - Information about the template alias.

$sel:status:DescribeTemplateAliasResponse', describeTemplateAliasResponse_status - The HTTP status of the request.

Response Lenses

describeTemplateAliasResponse_requestId :: Lens' DescribeTemplateAliasResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

describeTemplateAliasResponse_templateAlias :: Lens' DescribeTemplateAliasResponse (Maybe TemplateAlias)

Information about the template alias.

describeTemplateAliasResponse_status :: Lens' DescribeTemplateAliasResponse Int

The HTTP status of the request.