Amazonka.QuickSight.DescribeTheme

Description

Describes a theme.

Creating a Request

data DescribeTheme

See: newDescribeTheme smart constructor.

Constructors

DescribeTheme'

Fields

aliasName :: Maybe Text

The alias of the theme that you want to describe. If you name a specific alias, you describe the version that the alias points to. You can specify the latest version of the theme by providing the keyword $LATEST in the AliasName parameter. The keyword $PUBLISHED doesn't apply to themes.

versionNumber :: Maybe Natural

The version number for the version to describe. If a VersionNumber parameter value isn't provided, the latest version of the theme is described.

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the theme that you're describing.

themeId :: Text

The ID for the theme.

newDescribeTheme

Arguments

:: Text
$sel:awsAccountId:DescribeTheme'

-> Text
$sel:themeId:DescribeTheme'

-> DescribeTheme

Create a value of DescribeTheme with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aliasName:DescribeTheme', describeTheme_aliasName - The alias of the theme that you want to describe. If you name a specific alias, you describe the version that the alias points to. You can specify the latest version of the theme by providing the keyword $LATEST in the AliasName parameter. The keyword $PUBLISHED doesn't apply to themes.

$sel:versionNumber:DescribeTheme', describeTheme_versionNumber - The version number for the version to describe. If a VersionNumber parameter value isn't provided, the latest version of the theme is described.

$sel:awsAccountId:DescribeTheme', describeTheme_awsAccountId - The ID of the Amazon Web Services account that contains the theme that you're describing.

$sel:themeId:DescribeTheme', describeTheme_themeId - The ID for the theme.

Request Lenses

describeTheme_aliasName :: Lens' DescribeTheme (Maybe Text)

The alias of the theme that you want to describe. If you name a specific alias, you describe the version that the alias points to. You can specify the latest version of the theme by providing the keyword $LATEST in the AliasName parameter. The keyword $PUBLISHED doesn't apply to themes.

describeTheme_versionNumber :: Lens' DescribeTheme (Maybe Natural)

The version number for the version to describe. If a VersionNumber parameter value isn't provided, the latest version of the theme is described.

describeTheme_awsAccountId :: Lens' DescribeTheme Text

The ID of the Amazon Web Services account that contains the theme that you're describing.

describeTheme_themeId :: Lens' DescribeTheme Text

The ID for the theme.

Destructuring the Response

data DescribeThemeResponse

See: newDescribeThemeResponse smart constructor.

Constructors

DescribeThemeResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

theme :: Maybe Theme

The information about the theme that you are describing.

status :: Int

The HTTP status of the request.

newDescribeThemeResponse

Arguments

:: Int
$sel:status:DescribeThemeResponse'

-> DescribeThemeResponse

Create a value of DescribeThemeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DescribeThemeResponse', describeThemeResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:theme:DescribeThemeResponse', describeThemeResponse_theme - The information about the theme that you are describing.

$sel:status:DescribeThemeResponse', describeThemeResponse_status - The HTTP status of the request.

Response Lenses

describeThemeResponse_requestId :: Lens' DescribeThemeResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

describeThemeResponse_theme :: Lens' DescribeThemeResponse (Maybe Theme)

The information about the theme that you are describing.

describeThemeResponse_status :: Lens' DescribeThemeResponse Int

The HTTP status of the request.