Amazonka.QuickSight.DescribeThemeAlias

Description

Describes the alias for a theme.

Creating a Request

data DescribeThemeAlias

See: newDescribeThemeAlias smart constructor.

Constructors

DescribeThemeAlias'

Fields

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the theme alias that you're describing.

themeId :: Text

The ID for the theme.

aliasName :: Text

The name of the theme alias that you want to describe.

newDescribeThemeAlias

Arguments

:: Text
$sel:awsAccountId:DescribeThemeAlias'

-> Text
$sel:themeId:DescribeThemeAlias'

-> Text
$sel:aliasName:DescribeThemeAlias'

-> DescribeThemeAlias

Create a value of DescribeThemeAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:DescribeThemeAlias', describeThemeAlias_awsAccountId - The ID of the Amazon Web Services account that contains the theme alias that you're describing.

$sel:themeId:DescribeThemeAlias', describeThemeAlias_themeId - The ID for the theme.

$sel:aliasName:DescribeThemeAlias', describeThemeAlias_aliasName - The name of the theme alias that you want to describe.

Request Lenses

describeThemeAlias_awsAccountId :: Lens' DescribeThemeAlias Text

The ID of the Amazon Web Services account that contains the theme alias that you're describing.

describeThemeAlias_themeId :: Lens' DescribeThemeAlias Text

The ID for the theme.

describeThemeAlias_aliasName :: Lens' DescribeThemeAlias Text

The name of the theme alias that you want to describe.

Destructuring the Response

data DescribeThemeAliasResponse

See: newDescribeThemeAliasResponse smart constructor.

Constructors

DescribeThemeAliasResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

themeAlias :: Maybe ThemeAlias

Information about the theme alias.

status :: Int

The HTTP status of the request.

newDescribeThemeAliasResponse

Arguments

:: Int
$sel:status:DescribeThemeAliasResponse'

-> DescribeThemeAliasResponse

Create a value of DescribeThemeAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DescribeThemeAliasResponse', describeThemeAliasResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:themeAlias:DescribeThemeAliasResponse', describeThemeAliasResponse_themeAlias - Information about the theme alias.

$sel:status:DescribeThemeAliasResponse', describeThemeAliasResponse_status - The HTTP status of the request.

Response Lenses

describeThemeAliasResponse_requestId :: Lens' DescribeThemeAliasResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

describeThemeAliasResponse_themeAlias :: Lens' DescribeThemeAliasResponse (Maybe ThemeAlias)

Information about the theme alias.

describeThemeAliasResponse_status :: Lens' DescribeThemeAliasResponse Int

The HTTP status of the request.