Amazonka.QuickSight.GenerateEmbedUrlForAnonymousUser

Description

Generates an embed URL that you can use to embed an Amazon QuickSight dashboard in your website, without having to register any reader users. Before you use this action, make sure that you have configured the dashboards and permissions.

The following rules apply to the generated URL:

It contains a temporary bearer token. It is valid for 5 minutes after it is generated. Once redeemed within this period, it cannot be re-used again.

The URL validity period should not be confused with the actual session lifetime that can be customized using the SessionLifetimeInMinutes parameter.

The resulting user session is valid for 15 minutes (default) to 10 hours (maximum).

You are charged only when the URL is used or there is interaction with Amazon QuickSight.

For more information, see Embedded Analytics in the Amazon QuickSight User Guide.

For more information about the high-level steps for embedding and for an interactive demo of the ways you can customize embedding, visit the Amazon QuickSight Developer Portal.

Creating a Request

data GenerateEmbedUrlForAnonymousUser

See: newGenerateEmbedUrlForAnonymousUser smart constructor.

Constructors

GenerateEmbedUrlForAnonymousUser'

Fields

sessionLifetimeInMinutes :: Maybe Natural

How many minutes the session is valid. The session lifetime must be in [15-600] minutes range.

sessionTags :: Maybe (NonEmpty SessionTag)

The session tags used for row-level security. Before you use this parameter, make sure that you have configured the relevant datasets using the DataSet$RowLevelPermissionTagConfiguration parameter so that session tags can be used to provide row-level security.

These are not the tags used for the Amazon Web Services resource tagging feature. For more information, see Using Row-Level Security (RLS) with Tags.

awsAccountId :: Text

The ID for the Amazon Web Services account that contains the dashboard that you're embedding.

namespace :: Text

The Amazon QuickSight namespace that the anonymous user virtually belongs to. If you are not using an Amazon QuickSight custom namespace, set this to default.

authorizedResourceArns :: [Text]

The Amazon Resource Names for the Amazon QuickSight resources that the user is authorized to access during the lifetime of the session. If you choose Dashboard embedding experience, pass the list of dashboard ARNs in the account that you want the user to be able to view.

experienceConfiguration :: AnonymousUserEmbeddingExperienceConfiguration

The configuration of the experience you are embedding.

newGenerateEmbedUrlForAnonymousUser

Arguments

:: Text
$sel:awsAccountId:GenerateEmbedUrlForAnonymousUser'

-> Text
$sel:namespace:GenerateEmbedUrlForAnonymousUser'

-> AnonymousUserEmbeddingExperienceConfiguration
$sel:experienceConfiguration:GenerateEmbedUrlForAnonymousUser'

-> GenerateEmbedUrlForAnonymousUser

Create a value of GenerateEmbedUrlForAnonymousUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionLifetimeInMinutes:GenerateEmbedUrlForAnonymousUser', generateEmbedUrlForAnonymousUser_sessionLifetimeInMinutes - How many minutes the session is valid. The session lifetime must be in [15-600] minutes range.

$sel:sessionTags:GenerateEmbedUrlForAnonymousUser', generateEmbedUrlForAnonymousUser_sessionTags - The session tags used for row-level security. Before you use this parameter, make sure that you have configured the relevant datasets using the DataSet$RowLevelPermissionTagConfiguration parameter so that session tags can be used to provide row-level security.

These are not the tags used for the Amazon Web Services resource tagging feature. For more information, see Using Row-Level Security (RLS) with Tags.

$sel:awsAccountId:GenerateEmbedUrlForAnonymousUser', generateEmbedUrlForAnonymousUser_awsAccountId - The ID for the Amazon Web Services account that contains the dashboard that you're embedding.

$sel:namespace:GenerateEmbedUrlForAnonymousUser', generateEmbedUrlForAnonymousUser_namespace - The Amazon QuickSight namespace that the anonymous user virtually belongs to. If you are not using an Amazon QuickSight custom namespace, set this to default.

$sel:authorizedResourceArns:GenerateEmbedUrlForAnonymousUser', generateEmbedUrlForAnonymousUser_authorizedResourceArns - The Amazon Resource Names for the Amazon QuickSight resources that the user is authorized to access during the lifetime of the session. If you choose Dashboard embedding experience, pass the list of dashboard ARNs in the account that you want the user to be able to view.

$sel:experienceConfiguration:GenerateEmbedUrlForAnonymousUser', generateEmbedUrlForAnonymousUser_experienceConfiguration - The configuration of the experience you are embedding.

Request Lenses

generateEmbedUrlForAnonymousUser_sessionLifetimeInMinutes :: Lens' GenerateEmbedUrlForAnonymousUser (Maybe Natural)

How many minutes the session is valid. The session lifetime must be in [15-600] minutes range.

generateEmbedUrlForAnonymousUser_sessionTags :: Lens' GenerateEmbedUrlForAnonymousUser (Maybe (NonEmpty SessionTag))

The session tags used for row-level security. Before you use this parameter, make sure that you have configured the relevant datasets using the DataSet$RowLevelPermissionTagConfiguration parameter so that session tags can be used to provide row-level security.

These are not the tags used for the Amazon Web Services resource tagging feature. For more information, see Using Row-Level Security (RLS) with Tags.

generateEmbedUrlForAnonymousUser_awsAccountId :: Lens' GenerateEmbedUrlForAnonymousUser Text

The ID for the Amazon Web Services account that contains the dashboard that you're embedding.

generateEmbedUrlForAnonymousUser_namespace :: Lens' GenerateEmbedUrlForAnonymousUser Text

The Amazon QuickSight namespace that the anonymous user virtually belongs to. If you are not using an Amazon QuickSight custom namespace, set this to default.

generateEmbedUrlForAnonymousUser_authorizedResourceArns :: Lens' GenerateEmbedUrlForAnonymousUser [Text]

The Amazon Resource Names for the Amazon QuickSight resources that the user is authorized to access during the lifetime of the session. If you choose Dashboard embedding experience, pass the list of dashboard ARNs in the account that you want the user to be able to view.

generateEmbedUrlForAnonymousUser_experienceConfiguration :: Lens' GenerateEmbedUrlForAnonymousUser AnonymousUserEmbeddingExperienceConfiguration

The configuration of the experience you are embedding.

Destructuring the Response

data GenerateEmbedUrlForAnonymousUserResponse

See: newGenerateEmbedUrlForAnonymousUserResponse smart constructor.

Constructors

GenerateEmbedUrlForAnonymousUserResponse'

Fields

status :: Int

The HTTP status of the request.

embedUrl :: Sensitive Text

The embed URL for the dashboard.

requestId :: Text

The Amazon Web Services request ID for this operation.

newGenerateEmbedUrlForAnonymousUserResponse

Arguments

:: Int
$sel:status:GenerateEmbedUrlForAnonymousUserResponse'

-> Text
$sel:embedUrl:GenerateEmbedUrlForAnonymousUserResponse'

-> Text
$sel:requestId:GenerateEmbedUrlForAnonymousUserResponse'

-> GenerateEmbedUrlForAnonymousUserResponse

Create a value of GenerateEmbedUrlForAnonymousUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GenerateEmbedUrlForAnonymousUserResponse', generateEmbedUrlForAnonymousUserResponse_status - The HTTP status of the request.

$sel:embedUrl:GenerateEmbedUrlForAnonymousUserResponse', generateEmbedUrlForAnonymousUserResponse_embedUrl - The embed URL for the dashboard.

$sel:requestId:GenerateEmbedUrlForAnonymousUserResponse', generateEmbedUrlForAnonymousUserResponse_requestId - The Amazon Web Services request ID for this operation.

Response Lenses

generateEmbedUrlForAnonymousUserResponse_status :: Lens' GenerateEmbedUrlForAnonymousUserResponse Int

The HTTP status of the request.

generateEmbedUrlForAnonymousUserResponse_embedUrl :: Lens' GenerateEmbedUrlForAnonymousUserResponse Text

The embed URL for the dashboard.

generateEmbedUrlForAnonymousUserResponse_requestId :: Lens' GenerateEmbedUrlForAnonymousUserResponse Text

The Amazon Web Services request ID for this operation.