Amazonka.QuickSight.GenerateEmbedUrlForRegisteredUser

Description

Generates an embed URL that you can use to embed an Amazon QuickSight experience in your website. This action can be used for any type of user registered in an Amazon QuickSight account. Before you use this action, make sure that you have configured the relevant Amazon QuickSight resource and permissions.

The following rules apply to the generated URL:

It contains a temporary bearer token. It is valid for 5 minutes after it is generated. Once redeemed within this period, it cannot be re-used again.

The URL validity period should not be confused with the actual session lifetime that can be customized using the SessionLifetimeInMinutes parameter.

The resulting user session is valid for 15 minutes (default) to 10 hours (maximum).

You are charged only when the URL is used or there is interaction with Amazon QuickSight.

For more information, see Embedded Analytics in the Amazon QuickSight User Guide.

For more information about the high-level steps for embedding and for an interactive demo of the ways you can customize embedding, visit the Amazon QuickSight Developer Portal.

Creating a Request

data GenerateEmbedUrlForRegisteredUser

See: newGenerateEmbedUrlForRegisteredUser smart constructor.

Constructors

GenerateEmbedUrlForRegisteredUser'

Fields

sessionLifetimeInMinutes :: Maybe Natural

How many minutes the session is valid. The session lifetime must be in [15-600] minutes range.

awsAccountId :: Text

The ID for the Amazon Web Services account that contains the dashboard that you're embedding.

userArn :: Text

The Amazon Resource Name for the registered user.

experienceConfiguration :: RegisteredUserEmbeddingExperienceConfiguration

The experience you are embedding. For registered users, you can embed Amazon QuickSight dashboards, the entire Amazon QuickSight console, or the Amazon QuickSight Q search bar.

newGenerateEmbedUrlForRegisteredUser

Arguments

:: Text
$sel:awsAccountId:GenerateEmbedUrlForRegisteredUser'

-> Text
$sel:userArn:GenerateEmbedUrlForRegisteredUser'

-> RegisteredUserEmbeddingExperienceConfiguration
$sel:experienceConfiguration:GenerateEmbedUrlForRegisteredUser'

-> GenerateEmbedUrlForRegisteredUser

Create a value of GenerateEmbedUrlForRegisteredUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionLifetimeInMinutes:GenerateEmbedUrlForRegisteredUser', generateEmbedUrlForRegisteredUser_sessionLifetimeInMinutes - How many minutes the session is valid. The session lifetime must be in [15-600] minutes range.

$sel:awsAccountId:GenerateEmbedUrlForRegisteredUser', generateEmbedUrlForRegisteredUser_awsAccountId - The ID for the Amazon Web Services account that contains the dashboard that you're embedding.

$sel:userArn:GenerateEmbedUrlForRegisteredUser', generateEmbedUrlForRegisteredUser_userArn - The Amazon Resource Name for the registered user.

$sel:experienceConfiguration:GenerateEmbedUrlForRegisteredUser', generateEmbedUrlForRegisteredUser_experienceConfiguration - The experience you are embedding. For registered users, you can embed Amazon QuickSight dashboards, the entire Amazon QuickSight console, or the Amazon QuickSight Q search bar.

Request Lenses

generateEmbedUrlForRegisteredUser_sessionLifetimeInMinutes :: Lens' GenerateEmbedUrlForRegisteredUser (Maybe Natural)

How many minutes the session is valid. The session lifetime must be in [15-600] minutes range.

generateEmbedUrlForRegisteredUser_awsAccountId :: Lens' GenerateEmbedUrlForRegisteredUser Text

The ID for the Amazon Web Services account that contains the dashboard that you're embedding.

generateEmbedUrlForRegisteredUser_userArn :: Lens' GenerateEmbedUrlForRegisteredUser Text

The Amazon Resource Name for the registered user.

generateEmbedUrlForRegisteredUser_experienceConfiguration :: Lens' GenerateEmbedUrlForRegisteredUser RegisteredUserEmbeddingExperienceConfiguration

The experience you are embedding. For registered users, you can embed Amazon QuickSight dashboards, the entire Amazon QuickSight console, or the Amazon QuickSight Q search bar.

Destructuring the Response

data GenerateEmbedUrlForRegisteredUserResponse

See: newGenerateEmbedUrlForRegisteredUserResponse smart constructor.

Constructors

GenerateEmbedUrlForRegisteredUserResponse'

Fields

status :: Int

The HTTP status of the request.

embedUrl :: Sensitive Text

The embed URL for the Amazon QuickSight dashboard, console, or Q search bar.

requestId :: Text

The Amazon Web Services request ID for this operation.

newGenerateEmbedUrlForRegisteredUserResponse

Arguments

:: Int
$sel:status:GenerateEmbedUrlForRegisteredUserResponse'

-> Text
$sel:embedUrl:GenerateEmbedUrlForRegisteredUserResponse'

-> Text
$sel:requestId:GenerateEmbedUrlForRegisteredUserResponse'

-> GenerateEmbedUrlForRegisteredUserResponse

Create a value of GenerateEmbedUrlForRegisteredUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GenerateEmbedUrlForRegisteredUserResponse', generateEmbedUrlForRegisteredUserResponse_status - The HTTP status of the request.

$sel:embedUrl:GenerateEmbedUrlForRegisteredUserResponse', generateEmbedUrlForRegisteredUserResponse_embedUrl - The embed URL for the Amazon QuickSight dashboard, console, or Q search bar.

$sel:requestId:GenerateEmbedUrlForRegisteredUserResponse', generateEmbedUrlForRegisteredUserResponse_requestId - The Amazon Web Services request ID for this operation.

Response Lenses

generateEmbedUrlForRegisteredUserResponse_status :: Lens' GenerateEmbedUrlForRegisteredUserResponse Int

The HTTP status of the request.

generateEmbedUrlForRegisteredUserResponse_embedUrl :: Lens' GenerateEmbedUrlForRegisteredUserResponse Text

The embed URL for the Amazon QuickSight dashboard, console, or Q search bar.

generateEmbedUrlForRegisteredUserResponse_requestId :: Lens' GenerateEmbedUrlForRegisteredUserResponse Text

The Amazon Web Services request ID for this operation.