Amazonka.QuickSight.GetDashboardEmbedUrl

Description

Generates a session URL and authorization code that you can use to embed an Amazon Amazon QuickSight read-only dashboard in your web server code. Before you use this command, make sure that you have configured the dashboards and permissions.

Currently, you can use GetDashboardEmbedURL only from the server, not from the user's browser. The following rules apply to the combination of URL and authorization code:

They must be used together.

They can be used one time only.

They are valid for 5 minutes after you run this command.

The resulting user session is valid for 10 hours.

For more information, see Embedding Analytics Using GetDashboardEmbedUrl in the Amazon QuickSight User Guide.

For more information about the high-level steps for embedding and for an interactive demo of the ways you can customize embedding, visit the Amazon QuickSight Developer Portal.

Creating a Request

data GetDashboardEmbedUrl

See: newGetDashboardEmbedUrl smart constructor.

Constructors

GetDashboardEmbedUrl'

Fields

sessionLifetimeInMinutes :: Maybe Natural

How many minutes the session is valid. The session lifetime must be 15-600 minutes.

statePersistenceEnabled :: Maybe Bool

Adds persistence of state for the user session in an embedded dashboard. Persistence applies to the sheet and the parameter settings. These are control settings that the dashboard subscriber (Amazon QuickSight reader) chooses while viewing the dashboard. If this is set to TRUE, the settings are the same when the subscriber reopens the same dashboard URL. The state is stored in Amazon QuickSight, not in a browser cookie. If this is set to FALSE, the state of the user session is not persisted. The default is FALSE.

namespace :: Maybe Text

The Amazon QuickSight namespace that contains the dashboard IDs in this request. If you're not using a custom namespace, set Namespace = default.

additionalDashboardIds :: Maybe (NonEmpty Text)

A list of one or more dashboard IDs that you want to add to a session that includes anonymous users. The IdentityType parameter must be set to ANONYMOUS for this to work, because other identity types authenticate as Amazon QuickSight or IAMusers. For example, if you set "--dashboard-id dash_id1 --dashboard-id dash_id2 dash_id3 identity-type ANONYMOUS", the session can access all three dashboards.

undoRedoDisabled :: Maybe Bool

Remove the undo/redo button on the embedded dashboard. The default is FALSE, which enables the undo/redo button.

userArn :: Maybe Text

The Amazon QuickSight user's Amazon Resource Name (ARN), for use with QUICKSIGHT identity type. You can use this for any Amazon QuickSight users in your account (readers, authors, or admins) authenticated as one of the following:

Active Directory (AD) users or group members

Invited nonfederated users

IAMusers and IAMrole-based sessions authenticated through Federated Single Sign-On using SAML, OpenID Connect, or IAMfederation.

Omit this parameter for users in the third group – IAMusers and IAM role-based sessions.

resetDisabled :: Maybe Bool

Remove the reset button on the embedded dashboard. The default is FALSE, which enables the reset button.

awsAccountId :: Text

The ID for the Amazon Web Services account that contains the dashboard that you're embedding.

dashboardId :: Text

The ID for the dashboard, also added to the Identity and Access Management (IAM) policy.

identityType :: EmbeddingIdentityType

The authentication method that the user uses to sign in.

newGetDashboardEmbedUrl

Arguments

:: Text
$sel:awsAccountId:GetDashboardEmbedUrl'

-> Text
$sel:dashboardId:GetDashboardEmbedUrl'

-> EmbeddingIdentityType
$sel:identityType:GetDashboardEmbedUrl'

-> GetDashboardEmbedUrl

Create a value of GetDashboardEmbedUrl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionLifetimeInMinutes:GetDashboardEmbedUrl', getDashboardEmbedUrl_sessionLifetimeInMinutes - How many minutes the session is valid. The session lifetime must be 15-600 minutes.

$sel:statePersistenceEnabled:GetDashboardEmbedUrl', getDashboardEmbedUrl_statePersistenceEnabled - Adds persistence of state for the user session in an embedded dashboard. Persistence applies to the sheet and the parameter settings. These are control settings that the dashboard subscriber (Amazon QuickSight reader) chooses while viewing the dashboard. If this is set to TRUE, the settings are the same when the subscriber reopens the same dashboard URL. The state is stored in Amazon QuickSight, not in a browser cookie. If this is set to FALSE, the state of the user session is not persisted. The default is FALSE.

$sel:namespace:GetDashboardEmbedUrl', getDashboardEmbedUrl_namespace - The Amazon QuickSight namespace that contains the dashboard IDs in this request. If you're not using a custom namespace, set Namespace = default.

$sel:additionalDashboardIds:GetDashboardEmbedUrl', getDashboardEmbedUrl_additionalDashboardIds - A list of one or more dashboard IDs that you want to add to a session that includes anonymous users. The IdentityType parameter must be set to ANONYMOUS for this to work, because other identity types authenticate as Amazon QuickSight or IAMusers. For example, if you set "--dashboard-id dash_id1 --dashboard-id dash_id2 dash_id3 identity-type ANONYMOUS", the session can access all three dashboards.

$sel:undoRedoDisabled:GetDashboardEmbedUrl', getDashboardEmbedUrl_undoRedoDisabled - Remove the undo/redo button on the embedded dashboard. The default is FALSE, which enables the undo/redo button.

$sel:userArn:GetDashboardEmbedUrl', getDashboardEmbedUrl_userArn - The Amazon QuickSight user's Amazon Resource Name (ARN), for use with QUICKSIGHT identity type. You can use this for any Amazon QuickSight users in your account (readers, authors, or admins) authenticated as one of the following:

Active Directory (AD) users or group members

Invited nonfederated users

IAMusers and IAMrole-based sessions authenticated through Federated Single Sign-On using SAML, OpenID Connect, or IAMfederation.

Omit this parameter for users in the third group – IAMusers and IAM role-based sessions.

$sel:resetDisabled:GetDashboardEmbedUrl', getDashboardEmbedUrl_resetDisabled - Remove the reset button on the embedded dashboard. The default is FALSE, which enables the reset button.

$sel:awsAccountId:GetDashboardEmbedUrl', getDashboardEmbedUrl_awsAccountId - The ID for the Amazon Web Services account that contains the dashboard that you're embedding.

$sel:dashboardId:GetDashboardEmbedUrl', getDashboardEmbedUrl_dashboardId - The ID for the dashboard, also added to the Identity and Access Management (IAM) policy.

$sel:identityType:GetDashboardEmbedUrl', getDashboardEmbedUrl_identityType - The authentication method that the user uses to sign in.

Request Lenses

getDashboardEmbedUrl_sessionLifetimeInMinutes :: Lens' GetDashboardEmbedUrl (Maybe Natural)

How many minutes the session is valid. The session lifetime must be 15-600 minutes.

getDashboardEmbedUrl_statePersistenceEnabled :: Lens' GetDashboardEmbedUrl (Maybe Bool)

Adds persistence of state for the user session in an embedded dashboard. Persistence applies to the sheet and the parameter settings. These are control settings that the dashboard subscriber (Amazon QuickSight reader) chooses while viewing the dashboard. If this is set to TRUE, the settings are the same when the subscriber reopens the same dashboard URL. The state is stored in Amazon QuickSight, not in a browser cookie. If this is set to FALSE, the state of the user session is not persisted. The default is FALSE.

getDashboardEmbedUrl_namespace :: Lens' GetDashboardEmbedUrl (Maybe Text)

The Amazon QuickSight namespace that contains the dashboard IDs in this request. If you're not using a custom namespace, set Namespace = default.

getDashboardEmbedUrl_additionalDashboardIds :: Lens' GetDashboardEmbedUrl (Maybe (NonEmpty Text))

A list of one or more dashboard IDs that you want to add to a session that includes anonymous users. The IdentityType parameter must be set to ANONYMOUS for this to work, because other identity types authenticate as Amazon QuickSight or IAMusers. For example, if you set "--dashboard-id dash_id1 --dashboard-id dash_id2 dash_id3 identity-type ANONYMOUS", the session can access all three dashboards.

getDashboardEmbedUrl_undoRedoDisabled :: Lens' GetDashboardEmbedUrl (Maybe Bool)

Remove the undo/redo button on the embedded dashboard. The default is FALSE, which enables the undo/redo button.

getDashboardEmbedUrl_userArn :: Lens' GetDashboardEmbedUrl (Maybe Text)

The Amazon QuickSight user's Amazon Resource Name (ARN), for use with QUICKSIGHT identity type. You can use this for any Amazon QuickSight users in your account (readers, authors, or admins) authenticated as one of the following:

Active Directory (AD) users or group members

Invited nonfederated users

IAMusers and IAMrole-based sessions authenticated through Federated Single Sign-On using SAML, OpenID Connect, or IAMfederation.

Omit this parameter for users in the third group – IAMusers and IAM role-based sessions.

getDashboardEmbedUrl_resetDisabled :: Lens' GetDashboardEmbedUrl (Maybe Bool)

Remove the reset button on the embedded dashboard. The default is FALSE, which enables the reset button.

getDashboardEmbedUrl_awsAccountId :: Lens' GetDashboardEmbedUrl Text

The ID for the Amazon Web Services account that contains the dashboard that you're embedding.

getDashboardEmbedUrl_dashboardId :: Lens' GetDashboardEmbedUrl Text

The ID for the dashboard, also added to the Identity and Access Management (IAM) policy.

getDashboardEmbedUrl_identityType :: Lens' GetDashboardEmbedUrl EmbeddingIdentityType

The authentication method that the user uses to sign in.

Destructuring the Response

data GetDashboardEmbedUrlResponse

Output returned from the GetDashboardEmbedUrl operation.

See: newGetDashboardEmbedUrlResponse smart constructor.

Constructors

GetDashboardEmbedUrlResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

embedUrl :: Maybe (Sensitive Text)

A single-use URL that you can put into your server-side webpage to embed your dashboard. This URL is valid for 5 minutes. The API operation provides the URL with an auth_code value that enables one (and only one) sign-on to a user session that is valid for 10 hours.

status :: Int

The HTTP status of the request.

newGetDashboardEmbedUrlResponse

Arguments

:: Int
$sel:status:GetDashboardEmbedUrlResponse'

-> GetDashboardEmbedUrlResponse

Create a value of GetDashboardEmbedUrlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:GetDashboardEmbedUrlResponse', getDashboardEmbedUrlResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:embedUrl:GetDashboardEmbedUrlResponse', getDashboardEmbedUrlResponse_embedUrl - A single-use URL that you can put into your server-side webpage to embed your dashboard. This URL is valid for 5 minutes. The API operation provides the URL with an auth_code value that enables one (and only one) sign-on to a user session that is valid for 10 hours.

$sel:status:GetDashboardEmbedUrlResponse', getDashboardEmbedUrlResponse_status - The HTTP status of the request.

Response Lenses

getDashboardEmbedUrlResponse_requestId :: Lens' GetDashboardEmbedUrlResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

getDashboardEmbedUrlResponse_embedUrl :: Lens' GetDashboardEmbedUrlResponse (Maybe Text)

A single-use URL that you can put into your server-side webpage to embed your dashboard. This URL is valid for 5 minutes. The API operation provides the URL with an auth_code value that enables one (and only one) sign-on to a user session that is valid for 10 hours.

getDashboardEmbedUrlResponse_status :: Lens' GetDashboardEmbedUrlResponse Int

The HTTP status of the request.