Amazonka.QuickSight.ListAnalyses

Description

Lists Amazon QuickSight analyses that exist in the specified Amazon Web Services account.

This operation returns paginated results.

Creating a Request

data ListAnalyses

See: newListAnalyses smart constructor.

Constructors

ListAnalyses'

Fields

nextToken :: Maybe Text

A pagination token that can be used in a subsequent request.

maxResults :: Maybe Natural

The maximum number of results to return.

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the analyses.

newListAnalyses

Arguments

:: Text
$sel:awsAccountId:ListAnalyses'

-> ListAnalyses

Create a value of ListAnalyses with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAnalyses', listAnalyses_nextToken - A pagination token that can be used in a subsequent request.

$sel:maxResults:ListAnalyses', listAnalyses_maxResults - The maximum number of results to return.

$sel:awsAccountId:ListAnalyses', listAnalyses_awsAccountId - The ID of the Amazon Web Services account that contains the analyses.

Request Lenses

listAnalyses_nextToken :: Lens' ListAnalyses (Maybe Text)

A pagination token that can be used in a subsequent request.

listAnalyses_maxResults :: Lens' ListAnalyses (Maybe Natural)

The maximum number of results to return.

listAnalyses_awsAccountId :: Lens' ListAnalyses Text

The ID of the Amazon Web Services account that contains the analyses.

Destructuring the Response

data ListAnalysesResponse

See: newListAnalysesResponse smart constructor.

Constructors

ListAnalysesResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

nextToken :: Maybe Text

A pagination token that can be used in a subsequent request.

analysisSummaryList :: Maybe [AnalysisSummary]

Metadata describing each of the analyses that are listed.

status :: Int

The HTTP status of the request.

newListAnalysesResponse

Arguments

:: Int
$sel:status:ListAnalysesResponse'

-> ListAnalysesResponse

Create a value of ListAnalysesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:ListAnalysesResponse', listAnalysesResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:nextToken:ListAnalyses', listAnalysesResponse_nextToken - A pagination token that can be used in a subsequent request.

$sel:analysisSummaryList:ListAnalysesResponse', listAnalysesResponse_analysisSummaryList - Metadata describing each of the analyses that are listed.

$sel:status:ListAnalysesResponse', listAnalysesResponse_status - The HTTP status of the request.

Response Lenses

listAnalysesResponse_requestId :: Lens' ListAnalysesResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

listAnalysesResponse_nextToken :: Lens' ListAnalysesResponse (Maybe Text)

A pagination token that can be used in a subsequent request.

listAnalysesResponse_analysisSummaryList :: Lens' ListAnalysesResponse (Maybe [AnalysisSummary])

Metadata describing each of the analyses that are listed.

listAnalysesResponse_status :: Lens' ListAnalysesResponse Int

The HTTP status of the request.