Amazonka.QuickSight.ListIngestions

Description

Lists the history of SPICE ingestions for a dataset.

This operation returns paginated results.

Creating a Request

data ListIngestions

See: newListIngestions smart constructor.

Constructors

ListIngestions'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

maxResults :: Maybe Natural

The maximum number of results to be returned per request.

dataSetId :: Text

The ID of the dataset used in the ingestion.

awsAccountId :: Text

The Amazon Web Services account ID.

newListIngestions

Arguments

:: Text
$sel:dataSetId:ListIngestions'

-> Text
$sel:awsAccountId:ListIngestions'

-> ListIngestions

Create a value of ListIngestions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListIngestions', listIngestions_nextToken - The token for the next set of results, or null if there are no more results.

$sel:maxResults:ListIngestions', listIngestions_maxResults - The maximum number of results to be returned per request.

$sel:dataSetId:ListIngestions', listIngestions_dataSetId - The ID of the dataset used in the ingestion.

$sel:awsAccountId:ListIngestions', listIngestions_awsAccountId - The Amazon Web Services account ID.

Request Lenses

listIngestions_nextToken :: Lens' ListIngestions (Maybe Text)

The token for the next set of results, or null if there are no more results.

listIngestions_maxResults :: Lens' ListIngestions (Maybe Natural)

The maximum number of results to be returned per request.

listIngestions_dataSetId :: Lens' ListIngestions Text

The ID of the dataset used in the ingestion.

listIngestions_awsAccountId :: Lens' ListIngestions Text

The Amazon Web Services account ID.

Destructuring the Response

data ListIngestionsResponse

See: newListIngestionsResponse smart constructor.

Constructors

ListIngestionsResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

ingestions :: Maybe [Ingestion]

A list of the ingestions.

status :: Int

The HTTP status of the request.

newListIngestionsResponse

Arguments

:: Int
$sel:status:ListIngestionsResponse'

-> ListIngestionsResponse

Create a value of ListIngestionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:ListIngestionsResponse', listIngestionsResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:nextToken:ListIngestions', listIngestionsResponse_nextToken - The token for the next set of results, or null if there are no more results.

$sel:ingestions:ListIngestionsResponse', listIngestionsResponse_ingestions - A list of the ingestions.

$sel:status:ListIngestionsResponse', listIngestionsResponse_status - The HTTP status of the request.

Response Lenses

listIngestionsResponse_requestId :: Lens' ListIngestionsResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

listIngestionsResponse_nextToken :: Lens' ListIngestionsResponse (Maybe Text)

The token for the next set of results, or null if there are no more results.

listIngestionsResponse_ingestions :: Lens' ListIngestionsResponse (Maybe [Ingestion])

A list of the ingestions.

listIngestionsResponse_status :: Lens' ListIngestionsResponse Int

The HTTP status of the request.