Amazonka.QuickSight.ListThemeVersions

Description

Lists all the versions of the themes in the current Amazon Web Services account.

This operation returns paginated results.

Creating a Request

data ListThemeVersions

See: newListThemeVersions smart constructor.

Constructors

ListThemeVersions'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

maxResults :: Maybe Natural

The maximum number of results to be returned per request.

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the themes that you're listing.

themeId :: Text

The ID for the theme.

newListThemeVersions

Arguments

:: Text
$sel:awsAccountId:ListThemeVersions'

-> Text
$sel:themeId:ListThemeVersions'

-> ListThemeVersions

Create a value of ListThemeVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListThemeVersions', listThemeVersions_nextToken - The token for the next set of results, or null if there are no more results.

$sel:maxResults:ListThemeVersions', listThemeVersions_maxResults - The maximum number of results to be returned per request.

$sel:awsAccountId:ListThemeVersions', listThemeVersions_awsAccountId - The ID of the Amazon Web Services account that contains the themes that you're listing.

$sel:themeId:ListThemeVersions', listThemeVersions_themeId - The ID for the theme.

Request Lenses

listThemeVersions_nextToken :: Lens' ListThemeVersions (Maybe Text)

The token for the next set of results, or null if there are no more results.

listThemeVersions_maxResults :: Lens' ListThemeVersions (Maybe Natural)

The maximum number of results to be returned per request.

listThemeVersions_awsAccountId :: Lens' ListThemeVersions Text

The ID of the Amazon Web Services account that contains the themes that you're listing.

listThemeVersions_themeId :: Lens' ListThemeVersions Text

The ID for the theme.

Destructuring the Response

data ListThemeVersionsResponse

See: newListThemeVersionsResponse smart constructor.

Constructors

ListThemeVersionsResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

themeVersionSummaryList :: Maybe [ThemeVersionSummary]

A structure containing a list of all the versions of the specified theme.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

status :: Int

The HTTP status of the request.

newListThemeVersionsResponse

Arguments

:: Int
$sel:status:ListThemeVersionsResponse'

-> ListThemeVersionsResponse

Create a value of ListThemeVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:ListThemeVersionsResponse', listThemeVersionsResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:themeVersionSummaryList:ListThemeVersionsResponse', listThemeVersionsResponse_themeVersionSummaryList - A structure containing a list of all the versions of the specified theme.

$sel:nextToken:ListThemeVersions', listThemeVersionsResponse_nextToken - The token for the next set of results, or null if there are no more results.

$sel:status:ListThemeVersionsResponse', listThemeVersionsResponse_status - The HTTP status of the request.

Response Lenses

listThemeVersionsResponse_requestId :: Lens' ListThemeVersionsResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

listThemeVersionsResponse_themeVersionSummaryList :: Lens' ListThemeVersionsResponse (Maybe [ThemeVersionSummary])

A structure containing a list of all the versions of the specified theme.

listThemeVersionsResponse_nextToken :: Lens' ListThemeVersionsResponse (Maybe Text)

The token for the next set of results, or null if there are no more results.

listThemeVersionsResponse_status :: Lens' ListThemeVersionsResponse Int

The HTTP status of the request.