Amazonka.QuickSight.SearchFolders

Description

Searches the subfolders in a folder.

Creating a Request

data SearchFolders

See: newSearchFolders smart constructor.

Constructors

SearchFolders'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

maxResults :: Maybe Natural

The maximum number of results to be returned per request.

awsAccountId :: Text

The AWS account ID.

filters :: [FolderSearchFilter]

The filters to apply to the search. Currently, you can search only by the parent folder ARN. For example, "Filters": [ { "Name": "PARENT_FOLDER_ARN", "Operator": "StringEquals", "Value": "arn:aws:quicksight:us-east-1:1:folder/folderId" } ].

newSearchFolders

Arguments

:: Text
$sel:awsAccountId:SearchFolders'

-> SearchFolders

Create a value of SearchFolders with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchFolders', searchFolders_nextToken - The token for the next set of results, or null if there are no more results.

$sel:maxResults:SearchFolders', searchFolders_maxResults - The maximum number of results to be returned per request.

$sel:awsAccountId:SearchFolders', searchFolders_awsAccountId - The AWS account ID.

$sel:filters:SearchFolders', searchFolders_filters - The filters to apply to the search. Currently, you can search only by the parent folder ARN. For example, "Filters": [ { "Name": "PARENT_FOLDER_ARN", "Operator": "StringEquals", "Value": "arn:aws:quicksight:us-east-1:1:folder/folderId" } ].

Request Lenses

searchFolders_nextToken :: Lens' SearchFolders (Maybe Text)

The token for the next set of results, or null if there are no more results.

searchFolders_maxResults :: Lens' SearchFolders (Maybe Natural)

The maximum number of results to be returned per request.

searchFolders_awsAccountId :: Lens' SearchFolders Text

The AWS account ID.

searchFolders_filters :: Lens' SearchFolders [FolderSearchFilter]

The filters to apply to the search. Currently, you can search only by the parent folder ARN. For example, "Filters": [ { "Name": "PARENT_FOLDER_ARN", "Operator": "StringEquals", "Value": "arn:aws:quicksight:us-east-1:1:folder/folderId" } ].

Destructuring the Response

data SearchFoldersResponse

See: newSearchFoldersResponse smart constructor.

Constructors

SearchFoldersResponse'

Fields

requestId :: Maybe Text

The request ID.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

folderSummaryList :: Maybe [FolderSummary]

A structure that contains all of the folders in your AWS account. This structure provides basic information about the folders.

status :: Int

The status. If succeeded, the status is SC_OK.

newSearchFoldersResponse

Arguments

:: Int
$sel:status:SearchFoldersResponse'

-> SearchFoldersResponse

Create a value of SearchFoldersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:SearchFoldersResponse', searchFoldersResponse_requestId - The request ID.

$sel:nextToken:SearchFolders', searchFoldersResponse_nextToken - The token for the next set of results, or null if there are no more results.

$sel:folderSummaryList:SearchFoldersResponse', searchFoldersResponse_folderSummaryList - A structure that contains all of the folders in your AWS account. This structure provides basic information about the folders.

$sel:status:SearchFoldersResponse', searchFoldersResponse_status - The status. If succeeded, the status is SC_OK.

Response Lenses

searchFoldersResponse_requestId :: Lens' SearchFoldersResponse (Maybe Text)

The request ID.

searchFoldersResponse_nextToken :: Lens' SearchFoldersResponse (Maybe Text)

The token for the next set of results, or null if there are no more results.

searchFoldersResponse_folderSummaryList :: Lens' SearchFoldersResponse (Maybe [FolderSummary])

A structure that contains all of the folders in your AWS account. This structure provides basic information about the folders.

searchFoldersResponse_status :: Lens' SearchFoldersResponse Int

The status. If succeeded, the status is SC_OK.