Amazonka.QuickSight.Types.AnalysisSearchFilter

Description

Documentation

data AnalysisSearchFilter

A filter that you apply when searching for one or more analyses.

See: newAnalysisSearchFilter smart constructor.

Constructors

AnalysisSearchFilter'

Fields

operator :: Maybe FilterOperator

The comparison operator that you want to use as a filter, for example "Operator": "StringEquals".

value :: Maybe Text

The value of the named item, in this case QUICKSIGHT_USER, that you want to use as a filter, for example "Value". An example is "arn:aws:quicksight:us-east-1:1:user/default/UserName1".

name :: Maybe AnalysisFilterAttribute

The name of the value that you want to use as a filter, for example "Name": "QUICKSIGHT_USER".

newAnalysisSearchFilter :: AnalysisSearchFilter

Create a value of AnalysisSearchFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operator:AnalysisSearchFilter', analysisSearchFilter_operator - The comparison operator that you want to use as a filter, for example "Operator": "StringEquals".

$sel:value:AnalysisSearchFilter', analysisSearchFilter_value - The value of the named item, in this case QUICKSIGHT_USER, that you want to use as a filter, for example "Value". An example is "arn:aws:quicksight:us-east-1:1:user/default/UserName1".

$sel:name:AnalysisSearchFilter', analysisSearchFilter_name - The name of the value that you want to use as a filter, for example "Name": "QUICKSIGHT_USER".

analysisSearchFilter_operator :: Lens' AnalysisSearchFilter (Maybe FilterOperator)

The comparison operator that you want to use as a filter, for example "Operator": "StringEquals".

analysisSearchFilter_value :: Lens' AnalysisSearchFilter (Maybe Text)

The value of the named item, in this case QUICKSIGHT_USER, that you want to use as a filter, for example "Value". An example is "arn:aws:quicksight:us-east-1:1:user/default/UserName1".

analysisSearchFilter_name :: Lens' AnalysisSearchFilter (Maybe AnalysisFilterAttribute)

The name of the value that you want to use as a filter, for example "Name": "QUICKSIGHT_USER".