Amazonka.QuickSight.Types.Dashboard

Description

Documentation

data Dashboard

Dashboard.

See: newDashboard smart constructor.

Constructors

Dashboard'

Fields

lastUpdatedTime :: Maybe POSIX

The last time that this dataset was updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the resource.

createdTime :: Maybe POSIX

The time that this dataset was created.

dashboardId :: Maybe Text

Dashboard ID.

lastPublishedTime :: Maybe POSIX

The last time that this dataset was published.

name :: Maybe Text

A display name for the dashboard.

version :: Maybe DashboardVersion

Version.

newDashboard :: Dashboard

Create a value of Dashboard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:Dashboard', dashboard_lastUpdatedTime - The last time that this dataset was updated.

$sel:arn:Dashboard', dashboard_arn - The Amazon Resource Name (ARN) of the resource.

$sel:createdTime:Dashboard', dashboard_createdTime - The time that this dataset was created.

$sel:dashboardId:Dashboard', dashboard_dashboardId - Dashboard ID.

$sel:lastPublishedTime:Dashboard', dashboard_lastPublishedTime - The last time that this dataset was published.

$sel:name:Dashboard', dashboard_name - A display name for the dashboard.

$sel:version:Dashboard', dashboard_version - Version.

dashboard_lastUpdatedTime :: Lens' Dashboard (Maybe UTCTime)

The last time that this dataset was updated.

dashboard_arn :: Lens' Dashboard (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

dashboard_createdTime :: Lens' Dashboard (Maybe UTCTime)

The time that this dataset was created.

dashboard_dashboardId :: Lens' Dashboard (Maybe Text)

Dashboard ID.

dashboard_lastPublishedTime :: Lens' Dashboard (Maybe UTCTime)

The last time that this dataset was published.

dashboard_name :: Lens' Dashboard (Maybe Text)

A display name for the dashboard.

dashboard_version :: Lens' Dashboard (Maybe DashboardVersion)

Version.