Amazonka.QuickSight.Types.DataSet

Description

Documentation

data DataSet

Dataset.

See: newDataSet smart constructor.

Constructors

DataSet'

Fields

fieldFolders :: Maybe (HashMap Text FieldFolder)

The folder that contains fields and nested subfolders for your dataset.

columnGroups :: Maybe (NonEmpty ColumnGroup)

Groupings of columns that work together in certain Amazon QuickSight features. Currently, only geospatial hierarchy is supported.

lastUpdatedTime :: Maybe POSIX

The last time that this dataset was updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the resource.

createdTime :: Maybe POSIX

The time that this dataset was created.

consumedSpiceCapacityInBytes :: Maybe Integer

The amount of SPICE capacity used by this dataset. This is 0 if the dataset isn't imported into SPICE.

importMode :: Maybe DataSetImportMode

A value that indicates whether you want to import the data into SPICE.

physicalTableMap :: Maybe (HashMap Text PhysicalTable)

Declares the physical tables that are available in the underlying data sources.

dataSetId :: Maybe Text

The ID of the dataset.

name :: Maybe Text

A display name for the dataset.

dataSetUsageConfiguration :: Maybe DataSetUsageConfiguration

The usage configuration to apply to child datasets that reference this dataset as a source.

outputColumns :: Maybe [OutputColumn]

The list of columns after all transforms. These columns are available in templates, analyses, and dashboards.

rowLevelPermissionTagConfiguration :: Maybe RowLevelPermissionTagConfiguration

The element you can use to define tags for row-level security.

rowLevelPermissionDataSet :: Maybe RowLevelPermissionDataSet

The row-level security configuration for the dataset.

columnLevelPermissionRules :: Maybe (NonEmpty ColumnLevelPermissionRule)

A set of one or more definitions of a ColumnLevelPermissionRule .

logicalTableMap :: Maybe (HashMap Text LogicalTable)

Configures the combination and transformation of the data from the physical tables.

newDataSet :: DataSet

Create a value of DataSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fieldFolders:DataSet', dataSet_fieldFolders - The folder that contains fields and nested subfolders for your dataset.

$sel:columnGroups:DataSet', dataSet_columnGroups - Groupings of columns that work together in certain Amazon QuickSight features. Currently, only geospatial hierarchy is supported.

$sel:lastUpdatedTime:DataSet', dataSet_lastUpdatedTime - The last time that this dataset was updated.

$sel:arn:DataSet', dataSet_arn - The Amazon Resource Name (ARN) of the resource.

$sel:createdTime:DataSet', dataSet_createdTime - The time that this dataset was created.

$sel:consumedSpiceCapacityInBytes:DataSet', dataSet_consumedSpiceCapacityInBytes - The amount of SPICE capacity used by this dataset. This is 0 if the dataset isn't imported into SPICE.

$sel:importMode:DataSet', dataSet_importMode - A value that indicates whether you want to import the data into SPICE.

$sel:physicalTableMap:DataSet', dataSet_physicalTableMap - Declares the physical tables that are available in the underlying data sources.

$sel:dataSetId:DataSet', dataSet_dataSetId - The ID of the dataset.

$sel:name:DataSet', dataSet_name - A display name for the dataset.

$sel:dataSetUsageConfiguration:DataSet', dataSet_dataSetUsageConfiguration - The usage configuration to apply to child datasets that reference this dataset as a source.

$sel:outputColumns:DataSet', dataSet_outputColumns - The list of columns after all transforms. These columns are available in templates, analyses, and dashboards.

$sel:rowLevelPermissionTagConfiguration:DataSet', dataSet_rowLevelPermissionTagConfiguration - The element you can use to define tags for row-level security.

$sel:rowLevelPermissionDataSet:DataSet', dataSet_rowLevelPermissionDataSet - The row-level security configuration for the dataset.

$sel:columnLevelPermissionRules:DataSet', dataSet_columnLevelPermissionRules - A set of one or more definitions of a ColumnLevelPermissionRule .

$sel:logicalTableMap:DataSet', dataSet_logicalTableMap - Configures the combination and transformation of the data from the physical tables.

dataSet_fieldFolders :: Lens' DataSet (Maybe (HashMap Text FieldFolder))

The folder that contains fields and nested subfolders for your dataset.

dataSet_columnGroups :: Lens' DataSet (Maybe (NonEmpty ColumnGroup))

Groupings of columns that work together in certain Amazon QuickSight features. Currently, only geospatial hierarchy is supported.

dataSet_lastUpdatedTime :: Lens' DataSet (Maybe UTCTime)

The last time that this dataset was updated.

dataSet_arn :: Lens' DataSet (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

dataSet_createdTime :: Lens' DataSet (Maybe UTCTime)

The time that this dataset was created.

dataSet_consumedSpiceCapacityInBytes :: Lens' DataSet (Maybe Integer)

The amount of SPICE capacity used by this dataset. This is 0 if the dataset isn't imported into SPICE.

dataSet_importMode :: Lens' DataSet (Maybe DataSetImportMode)

A value that indicates whether you want to import the data into SPICE.

dataSet_physicalTableMap :: Lens' DataSet (Maybe (HashMap Text PhysicalTable))

Declares the physical tables that are available in the underlying data sources.

dataSet_dataSetId :: Lens' DataSet (Maybe Text)

The ID of the dataset.

dataSet_name :: Lens' DataSet (Maybe Text)

A display name for the dataset.

dataSet_dataSetUsageConfiguration :: Lens' DataSet (Maybe DataSetUsageConfiguration)

The usage configuration to apply to child datasets that reference this dataset as a source.

dataSet_outputColumns :: Lens' DataSet (Maybe [OutputColumn])

The list of columns after all transforms. These columns are available in templates, analyses, and dashboards.

dataSet_rowLevelPermissionTagConfiguration :: Lens' DataSet (Maybe RowLevelPermissionTagConfiguration)

The element you can use to define tags for row-level security.

dataSet_rowLevelPermissionDataSet :: Lens' DataSet (Maybe RowLevelPermissionDataSet)

The row-level security configuration for the dataset.

dataSet_columnLevelPermissionRules :: Lens' DataSet (Maybe (NonEmpty ColumnLevelPermissionRule))

A set of one or more definitions of a ColumnLevelPermissionRule .

dataSet_logicalTableMap :: Lens' DataSet (Maybe (HashMap Text LogicalTable))

Configures the combination and transformation of the data from the physical tables.