Amazonka.QuickSight.Types.DataSetConfiguration

Description

Documentation

data DataSetConfiguration

Dataset configuration.

See: newDataSetConfiguration smart constructor.

Constructors

DataSetConfiguration'

Fields

columnGroupSchemaList :: Maybe [ColumnGroupSchema]

A structure containing the list of column group schemas.

dataSetSchema :: Maybe DataSetSchema

Dataset schema.

placeholder :: Maybe Text

Placeholder.

newDataSetConfiguration :: DataSetConfiguration

Create a value of DataSetConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:columnGroupSchemaList:DataSetConfiguration', dataSetConfiguration_columnGroupSchemaList - A structure containing the list of column group schemas.

$sel:dataSetSchema:DataSetConfiguration', dataSetConfiguration_dataSetSchema - Dataset schema.

$sel:placeholder:DataSetConfiguration', dataSetConfiguration_placeholder - Placeholder.

dataSetConfiguration_columnGroupSchemaList :: Lens' DataSetConfiguration (Maybe [ColumnGroupSchema])

A structure containing the list of column group schemas.

dataSetConfiguration_dataSetSchema :: Lens' DataSetConfiguration (Maybe DataSetSchema)

Dataset schema.

dataSetConfiguration_placeholder :: Lens' DataSetConfiguration (Maybe Text)

Placeholder.