Amazonka.QuickSight.Types.FolderMember

Description

Documentation

data FolderMember

An asset in a folder, such as a dashboard, analysis, or dataset.

See: newFolderMember smart constructor.

Constructors

FolderMember'

Fields

memberId :: Maybe Text

The ID of the asset.

memberType :: Maybe MemberType

The type of the asset.

newFolderMember :: FolderMember

Create a value of FolderMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberId:FolderMember', folderMember_memberId - The ID of the asset.

$sel:memberType:FolderMember', folderMember_memberType - The type of the asset.

folderMember_memberId :: Lens' FolderMember (Maybe Text)

The ID of the asset.

folderMember_memberType :: Lens' FolderMember (Maybe MemberType)

The type of the asset.