Amazonka.QuickSight.Types.FolderSearchFilter

Description

Documentation

data FolderSearchFilter

Searches a folder by a filter.

See: newFolderSearchFilter smart constructor.

Constructors

FolderSearchFilter'

Fields

operator :: Maybe FilterOperator

The comparison operator that you want to use as a filter. For example, "Operator": "StringEquals".

value :: Maybe Text

The value of the named item (in this example, PARENT_FOLDER_ARN), that you want to use as a filter. For example, "Value": "arn:aws:quicksight:us-east-1:1:folder/folderId".

name :: Maybe FolderFilterAttribute

The name of the value that you want to use as a filter. For example, "Name": "PARENT_FOLDER_ARN".

newFolderSearchFilter :: FolderSearchFilter

Create a value of FolderSearchFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operator:FolderSearchFilter', folderSearchFilter_operator - The comparison operator that you want to use as a filter. For example, "Operator": "StringEquals".

$sel:value:FolderSearchFilter', folderSearchFilter_value - The value of the named item (in this example, PARENT_FOLDER_ARN), that you want to use as a filter. For example, "Value": "arn:aws:quicksight:us-east-1:1:folder/folderId".

$sel:name:FolderSearchFilter', folderSearchFilter_name - The name of the value that you want to use as a filter. For example, "Name": "PARENT_FOLDER_ARN".

folderSearchFilter_operator :: Lens' FolderSearchFilter (Maybe FilterOperator)

The comparison operator that you want to use as a filter. For example, "Operator": "StringEquals".

folderSearchFilter_value :: Lens' FolderSearchFilter (Maybe Text)

The value of the named item (in this example, PARENT_FOLDER_ARN), that you want to use as a filter. For example, "Value": "arn:aws:quicksight:us-east-1:1:folder/folderId".

folderSearchFilter_name :: Lens' FolderSearchFilter (Maybe FolderFilterAttribute)

The name of the value that you want to use as a filter. For example, "Name": "PARENT_FOLDER_ARN".