Amazonka.QuickSight.Types.Ingestion

Description

Documentation

data Ingestion

Information about the SPICE ingestion for a dataset.

See: newIngestion smart constructor.

Constructors

Ingestion'

Fields

requestSource :: Maybe IngestionRequestSource

Event source for this ingestion.

queueInfo :: Maybe QueueInfo

ingestionTimeInSeconds :: Maybe Integer

The time that this ingestion took, measured in seconds.

requestType :: Maybe IngestionRequestType

Type of this ingestion.

ingestionSizeInBytes :: Maybe Integer

The size of the data ingested, in bytes.

rowInfo :: Maybe RowInfo

ingestionId :: Maybe Text

Ingestion ID.

errorInfo :: Maybe ErrorInfo

Error information for this ingestion.

arn :: Text

The Amazon Resource Name (ARN) of the resource.

ingestionStatus :: IngestionStatus

Ingestion status.

createdTime :: POSIX

The time that this ingestion started.

newIngestion

Arguments

:: Text
$sel:arn:Ingestion'

-> IngestionStatus
$sel:ingestionStatus:Ingestion'

-> UTCTime
$sel:createdTime:Ingestion'

-> Ingestion

Create a value of Ingestion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestSource:Ingestion', ingestion_requestSource - Event source for this ingestion.

$sel:queueInfo:Ingestion', ingestion_queueInfo - Undocumented member.

$sel:ingestionTimeInSeconds:Ingestion', ingestion_ingestionTimeInSeconds - The time that this ingestion took, measured in seconds.

$sel:requestType:Ingestion', ingestion_requestType - Type of this ingestion.

$sel:ingestionSizeInBytes:Ingestion', ingestion_ingestionSizeInBytes - The size of the data ingested, in bytes.

$sel:rowInfo:Ingestion', ingestion_rowInfo - Undocumented member.

$sel:ingestionId:Ingestion', ingestion_ingestionId - Ingestion ID.

$sel:errorInfo:Ingestion', ingestion_errorInfo - Error information for this ingestion.

$sel:arn:Ingestion', ingestion_arn - The Amazon Resource Name (ARN) of the resource.

$sel:ingestionStatus:Ingestion', ingestion_ingestionStatus - Ingestion status.

$sel:createdTime:Ingestion', ingestion_createdTime - The time that this ingestion started.

ingestion_requestSource :: Lens' Ingestion (Maybe IngestionRequestSource)

Event source for this ingestion.

ingestion_queueInfo :: Lens' Ingestion (Maybe QueueInfo)

Undocumented member.

ingestion_ingestionTimeInSeconds :: Lens' Ingestion (Maybe Integer)

The time that this ingestion took, measured in seconds.

ingestion_requestType :: Lens' Ingestion (Maybe IngestionRequestType)

Type of this ingestion.

ingestion_ingestionSizeInBytes :: Lens' Ingestion (Maybe Integer)

The size of the data ingested, in bytes.

ingestion_rowInfo :: Lens' Ingestion (Maybe RowInfo)

Undocumented member.

ingestion_ingestionId :: Lens' Ingestion (Maybe Text)

Ingestion ID.

ingestion_errorInfo :: Lens' Ingestion (Maybe ErrorInfo)

Error information for this ingestion.

ingestion_arn :: Lens' Ingestion Text

The Amazon Resource Name (ARN) of the resource.

ingestion_ingestionStatus :: Lens' Ingestion IngestionStatus

Ingestion status.

ingestion_createdTime :: Lens' Ingestion UTCTime

The time that this ingestion started.