Amazonka.QuickSight.Types.Parameters

Description

Documentation

data Parameters

A list of Amazon QuickSight parameters and the list's override values.

See: newParameters smart constructor.

Constructors

Parameters'

Fields

dateTimeParameters :: Maybe [DateTimeParameter]

The parameters that have a data type of date-time.

decimalParameters :: Maybe [DecimalParameter]

The parameters that have a data type of decimal.

integerParameters :: Maybe [IntegerParameter]

The parameters that have a data type of integer.

stringParameters :: Maybe [StringParameter]

The parameters that have a data type of string.

newParameters :: Parameters

Create a value of Parameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dateTimeParameters:Parameters', parameters_dateTimeParameters - The parameters that have a data type of date-time.

$sel:decimalParameters:Parameters', parameters_decimalParameters - The parameters that have a data type of decimal.

$sel:integerParameters:Parameters', parameters_integerParameters - The parameters that have a data type of integer.

$sel:stringParameters:Parameters', parameters_stringParameters - The parameters that have a data type of string.

parameters_dateTimeParameters :: Lens' Parameters (Maybe [DateTimeParameter])

The parameters that have a data type of date-time.

parameters_decimalParameters :: Lens' Parameters (Maybe [DecimalParameter])

The parameters that have a data type of decimal.

parameters_integerParameters :: Lens' Parameters (Maybe [IntegerParameter])

The parameters that have a data type of integer.

parameters_stringParameters :: Lens' Parameters (Maybe [StringParameter])

The parameters that have a data type of string.