Amazonka.QuickSight.Types.RegisteredUserEmbeddingExperienceConfiguration

Description

Documentation

data RegisteredUserEmbeddingExperienceConfiguration

The type of experience you want to embed. For registered users, you can embed an Amazon QuickSight dashboard or the Amazon QuickSight console.

Exactly one of the experience configurations is required. You can choose Dashboard or QuickSightConsole. You cannot choose more than one experience configuraton.

See: newRegisteredUserEmbeddingExperienceConfiguration smart constructor.

Constructors

RegisteredUserEmbeddingExperienceConfiguration'

Fields

dashboard :: Maybe RegisteredUserDashboardEmbeddingConfiguration

The configuration details for providing a dashboard embedding experience.

quickSightConsole :: Maybe RegisteredUserQuickSightConsoleEmbeddingConfiguration

The configuration details for providing an Amazon QuickSight console embedding experience. This can be used along with custom permissions to restrict access to certain features. For more information, see Customizing Access to the Amazon QuickSight Console in the Amazon QuickSight User Guide.

Use GenerateEmbedUrlForRegisteredUser where you want to provide an authoring portal that allows users to create data sources, datasets, analyses, and dashboards. The users who accesses an embedded Amazon QuickSight console needs to belong to the author or admin security cohort. If you want to restrict permissions to some of these features, add a custom permissions profile to the user with the UpdateUser API operation. Use RegisterUser API operation to add a new user with a custom permission profile attached. For more information, see the following sections in the Amazon QuickSight User Guide:

Embedding the Full Functionality of the Amazon QuickSight Console for Authenticated Users

Customizing Access to the Amazon QuickSight Console

For more information about the high-level steps for embedding and for an interactive demo of the ways you can customize embedding, visit the Amazon QuickSight Developer Portal.

qSearchBar :: Maybe RegisteredUserQSearchBarEmbeddingConfiguration

The configuration details for embedding the Q search bar.

For more information about embedding the Q search bar, see Embedding Overview.

newRegisteredUserEmbeddingExperienceConfiguration :: RegisteredUserEmbeddingExperienceConfiguration

Create a value of RegisteredUserEmbeddingExperienceConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dashboard:RegisteredUserEmbeddingExperienceConfiguration', registeredUserEmbeddingExperienceConfiguration_dashboard - The configuration details for providing a dashboard embedding experience.

$sel:quickSightConsole:RegisteredUserEmbeddingExperienceConfiguration', registeredUserEmbeddingExperienceConfiguration_quickSightConsole - The configuration details for providing an Amazon QuickSight console embedding experience. This can be used along with custom permissions to restrict access to certain features. For more information, see Customizing Access to the Amazon QuickSight Console in the Amazon QuickSight User Guide.

Use GenerateEmbedUrlForRegisteredUser where you want to provide an authoring portal that allows users to create data sources, datasets, analyses, and dashboards. The users who accesses an embedded Amazon QuickSight console needs to belong to the author or admin security cohort. If you want to restrict permissions to some of these features, add a custom permissions profile to the user with the UpdateUser API operation. Use RegisterUser API operation to add a new user with a custom permission profile attached. For more information, see the following sections in the Amazon QuickSight User Guide:

Embedding the Full Functionality of the Amazon QuickSight Console for Authenticated Users

Customizing Access to the Amazon QuickSight Console

For more information about the high-level steps for embedding and for an interactive demo of the ways you can customize embedding, visit the Amazon QuickSight Developer Portal.

$sel:qSearchBar:RegisteredUserEmbeddingExperienceConfiguration', registeredUserEmbeddingExperienceConfiguration_qSearchBar - The configuration details for embedding the Q search bar.

For more information about embedding the Q search bar, see Embedding Overview.

registeredUserEmbeddingExperienceConfiguration_dashboard :: Lens' RegisteredUserEmbeddingExperienceConfiguration (Maybe RegisteredUserDashboardEmbeddingConfiguration)

The configuration details for providing a dashboard embedding experience.

registeredUserEmbeddingExperienceConfiguration_quickSightConsole :: Lens' RegisteredUserEmbeddingExperienceConfiguration (Maybe RegisteredUserQuickSightConsoleEmbeddingConfiguration)

The configuration details for providing an Amazon QuickSight console embedding experience. This can be used along with custom permissions to restrict access to certain features. For more information, see Customizing Access to the Amazon QuickSight Console in the Amazon QuickSight User Guide.

Use GenerateEmbedUrlForRegisteredUser where you want to provide an authoring portal that allows users to create data sources, datasets, analyses, and dashboards. The users who accesses an embedded Amazon QuickSight console needs to belong to the author or admin security cohort. If you want to restrict permissions to some of these features, add a custom permissions profile to the user with the UpdateUser API operation. Use RegisterUser API operation to add a new user with a custom permission profile attached. For more information, see the following sections in the Amazon QuickSight User Guide:

Embedding the Full Functionality of the Amazon QuickSight Console for Authenticated Users

Customizing Access to the Amazon QuickSight Console

For more information about the high-level steps for embedding and for an interactive demo of the ways you can customize embedding, visit the Amazon QuickSight Developer Portal.

registeredUserEmbeddingExperienceConfiguration_qSearchBar :: Lens' RegisteredUserEmbeddingExperienceConfiguration (Maybe RegisteredUserQSearchBarEmbeddingConfiguration)

The configuration details for embedding the Q search bar.

For more information about embedding the Q search bar, see Embedding Overview.