Amazonka.QuickSight.Types.ResourcePermission

Description

Documentation

data ResourcePermission

Permission for the resource.

See: newResourcePermission smart constructor.

Constructors

ResourcePermission'

Fields

principal :: Text

The Amazon Resource Name (ARN) of the principal. This can be one of the following:

The ARN of an Amazon QuickSight user or group associated with a data source or dataset. (This is common.)

The ARN of an Amazon QuickSight user, group, or namespace associated with an analysis, dashboard, template, or theme. (This is common.)

The ARN of an Amazon Web Services account root: This is an IAMARN rather than a Amazon QuickSight ARN. Use this option only to share resources (templates) across Amazon Web Services accounts. (This is less common.)

actions :: NonEmpty Text

The IAMaction to grant or revoke permissions on.

newResourcePermission

Arguments

:: Text
$sel:principal:ResourcePermission'

-> NonEmpty Text
$sel:actions:ResourcePermission'

-> ResourcePermission

Create a value of ResourcePermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principal:ResourcePermission', resourcePermission_principal - The Amazon Resource Name (ARN) of the principal. This can be one of the following:

The ARN of an Amazon QuickSight user or group associated with a data source or dataset. (This is common.)

The ARN of an Amazon QuickSight user, group, or namespace associated with an analysis, dashboard, template, or theme. (This is common.)

The ARN of an Amazon Web Services account root: This is an IAMARN rather than a Amazon QuickSight ARN. Use this option only to share resources (templates) across Amazon Web Services accounts. (This is less common.)

$sel:actions:ResourcePermission', resourcePermission_actions - The IAMaction to grant or revoke permissions on.

resourcePermission_principal :: Lens' ResourcePermission Text

The Amazon Resource Name (ARN) of the principal. This can be one of the following:

The ARN of an Amazon QuickSight user or group associated with a data source or dataset. (This is common.)

The ARN of an Amazon QuickSight user, group, or namespace associated with an analysis, dashboard, template, or theme. (This is common.)

The ARN of an Amazon Web Services account root: This is an IAMARN rather than a Amazon QuickSight ARN. Use this option only to share resources (templates) across Amazon Web Services accounts. (This is less common.)

resourcePermission_actions :: Lens' ResourcePermission (NonEmpty Text)

The IAMaction to grant or revoke permissions on.