Amazonka.QuickSight.Types.RowLevelPermissionTagConfiguration

Description

Documentation

data RowLevelPermissionTagConfiguration

The configuration of tags on a dataset to set row-level security.

See: newRowLevelPermissionTagConfiguration smart constructor.

Constructors

RowLevelPermissionTagConfiguration'

Fields

status :: Maybe Status

The status of row-level security tags. If enabled, the status is ENABLED. If disabled, the status is DISABLED.

tagRules :: NonEmpty RowLevelPermissionTagRule

A set of rules associated with row-level security, such as the tag names and columns that they are assigned to.

newRowLevelPermissionTagConfiguration

Arguments

:: NonEmpty RowLevelPermissionTagRule
$sel:tagRules:RowLevelPermissionTagConfiguration'

-> RowLevelPermissionTagConfiguration

Create a value of RowLevelPermissionTagConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RowLevelPermissionTagConfiguration', rowLevelPermissionTagConfiguration_status - The status of row-level security tags. If enabled, the status is ENABLED. If disabled, the status is DISABLED.

$sel:tagRules:RowLevelPermissionTagConfiguration', rowLevelPermissionTagConfiguration_tagRules - A set of rules associated with row-level security, such as the tag names and columns that they are assigned to.

rowLevelPermissionTagConfiguration_status :: Lens' RowLevelPermissionTagConfiguration (Maybe Status)

The status of row-level security tags. If enabled, the status is ENABLED. If disabled, the status is DISABLED.

rowLevelPermissionTagConfiguration_tagRules :: Lens' RowLevelPermissionTagConfiguration (NonEmpty RowLevelPermissionTagRule)

A set of rules associated with row-level security, such as the tag names and columns that they are assigned to.