Amazonka.QuickSight.Types.TemplateVersion

Description

Documentation

data TemplateVersion

A version of a template.

See: newTemplateVersion smart constructor.

Constructors

TemplateVersion'

Fields

status :: Maybe ResourceStatus

The HTTP status of the request.

themeArn :: Maybe Text

The ARN of the theme associated with this version of the template.

sheets :: Maybe [Sheet]

A list of the associated sheets with the unique identifier and name of each sheet.

createdTime :: Maybe POSIX

The time that this template version was created.

sourceEntityArn :: Maybe Text

The Amazon Resource Name (ARN) of an analysis or template that was used to create this template.

dataSetConfigurations :: Maybe [DataSetConfiguration]

Schema of the dataset identified by the placeholder. Any dashboard created from this template should be bound to new datasets matching the same schema described through this API operation.

versionNumber :: Maybe Natural

The version number of the template version.

errors :: Maybe (NonEmpty TemplateError)

Errors associated with this template version.

description :: Maybe Text

The description of the template.

newTemplateVersion :: TemplateVersion

Create a value of TemplateVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:TemplateVersion', templateVersion_status - The HTTP status of the request.

$sel:themeArn:TemplateVersion', templateVersion_themeArn - The ARN of the theme associated with this version of the template.

$sel:sheets:TemplateVersion', templateVersion_sheets - A list of the associated sheets with the unique identifier and name of each sheet.

$sel:createdTime:TemplateVersion', templateVersion_createdTime - The time that this template version was created.

$sel:sourceEntityArn:TemplateVersion', templateVersion_sourceEntityArn - The Amazon Resource Name (ARN) of an analysis or template that was used to create this template.

$sel:dataSetConfigurations:TemplateVersion', templateVersion_dataSetConfigurations - Schema of the dataset identified by the placeholder. Any dashboard created from this template should be bound to new datasets matching the same schema described through this API operation.

$sel:versionNumber:TemplateVersion', templateVersion_versionNumber - The version number of the template version.

$sel:errors:TemplateVersion', templateVersion_errors - Errors associated with this template version.

$sel:description:TemplateVersion', templateVersion_description - The description of the template.

templateVersion_status :: Lens' TemplateVersion (Maybe ResourceStatus)

The HTTP status of the request.

templateVersion_themeArn :: Lens' TemplateVersion (Maybe Text)

The ARN of the theme associated with this version of the template.

templateVersion_sheets :: Lens' TemplateVersion (Maybe [Sheet])

A list of the associated sheets with the unique identifier and name of each sheet.

templateVersion_createdTime :: Lens' TemplateVersion (Maybe UTCTime)

The time that this template version was created.

templateVersion_sourceEntityArn :: Lens' TemplateVersion (Maybe Text)

The Amazon Resource Name (ARN) of an analysis or template that was used to create this template.

templateVersion_dataSetConfigurations :: Lens' TemplateVersion (Maybe [DataSetConfiguration])

Schema of the dataset identified by the placeholder. Any dashboard created from this template should be bound to new datasets matching the same schema described through this API operation.

templateVersion_versionNumber :: Lens' TemplateVersion (Maybe Natural)

The version number of the template version.

templateVersion_errors :: Lens' TemplateVersion (Maybe (NonEmpty TemplateError))

Errors associated with this template version.

templateVersion_description :: Lens' TemplateVersion (Maybe Text)

The description of the template.