Amazonka.QuickSight.Types.Theme

Description

Documentation

data Theme

Summary information about a theme.

See: newTheme smart constructor.

Constructors

Theme'

Fields

themeId :: Maybe Text

The identifier that the user gives to the theme.

lastUpdatedTime :: Maybe POSIX

The date and time that the theme was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the theme.

createdTime :: Maybe POSIX

The date and time that the theme was created.

name :: Maybe Text

The name that the user gives to the theme.

version :: Maybe ThemeVersion

type' :: Maybe ThemeType

The type of theme, based on how it was created. Valid values include: QUICKSIGHT and CUSTOM.

newTheme :: Theme

Create a value of Theme with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:themeId:Theme', theme_themeId - The identifier that the user gives to the theme.

$sel:lastUpdatedTime:Theme', theme_lastUpdatedTime - The date and time that the theme was last updated.

$sel:arn:Theme', theme_arn - The Amazon Resource Name (ARN) of the theme.

$sel:createdTime:Theme', theme_createdTime - The date and time that the theme was created.

$sel:name:Theme', theme_name - The name that the user gives to the theme.

$sel:version:Theme', theme_version - Undocumented member.

$sel:type':Theme', theme_type - The type of theme, based on how it was created. Valid values include: QUICKSIGHT and CUSTOM.

theme_themeId :: Lens' Theme (Maybe Text)

The identifier that the user gives to the theme.

theme_lastUpdatedTime :: Lens' Theme (Maybe UTCTime)

The date and time that the theme was last updated.

theme_arn :: Lens' Theme (Maybe Text)

The Amazon Resource Name (ARN) of the theme.

theme_createdTime :: Lens' Theme (Maybe UTCTime)

The date and time that the theme was created.

theme_name :: Lens' Theme (Maybe Text)

The name that the user gives to the theme.

theme_version :: Lens' Theme (Maybe ThemeVersion)

Undocumented member.

theme_type :: Lens' Theme (Maybe ThemeType)

The type of theme, based on how it was created. Valid values include: QUICKSIGHT and CUSTOM.