Amazonka.QuickSight.Types.UIColorPalette

Description

Documentation

data UIColorPalette

The theme colors that apply to UI and to charts, excluding data colors. The colors description is a hexadecimal color code that consists of six alphanumerical characters, prefixed with #, for example #37BFF5. For more information, see Using Themes in Amazon QuickSight in the Amazon QuickSight User Guide.

See: newUIColorPalette smart constructor.

Constructors

UIColorPalette'

Fields

secondaryBackground :: Maybe Text

The background color that applies to the sheet background and sheet controls.

successForeground :: Maybe Text

The foreground color that applies to any text or other elements that appear over the success color.

dimension :: Maybe Text

The color that applies to the names of fields that are identified as dimensions.

dimensionForeground :: Maybe Text

The foreground color that applies to any text or other elements that appear over the dimension color.

success :: Maybe Text

The color that applies to success messages, for example the check mark for a successful download.

accent :: Maybe Text

This color is that applies to selected states and buttons.

secondaryForeground :: Maybe Text

The foreground color that applies to any sheet title, sheet control text, or UI that appears over the secondary background.

primaryForeground :: Maybe Text

The color of text and other foreground elements that appear over the primary background regions, such as grid lines, borders, table banding, icons, and so on.

warning :: Maybe Text

This color that applies to warning and informational messages.

primaryBackground :: Maybe Text

The background color that applies to visuals and other high emphasis UI.

accentForeground :: Maybe Text

The foreground color that applies to any text or other elements that appear over the accent color.

measure :: Maybe Text

The color that applies to the names of fields that are identified as measures.

dangerForeground :: Maybe Text

The foreground color that applies to any text or other elements that appear over the error color.

measureForeground :: Maybe Text

The foreground color that applies to any text or other elements that appear over the measure color.

danger :: Maybe Text

The color that applies to error messages.

warningForeground :: Maybe Text

The foreground color that applies to any text or other elements that appear over the warning color.

newUIColorPalette :: UIColorPalette

Create a value of UIColorPalette with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secondaryBackground:UIColorPalette', uIColorPalette_secondaryBackground - The background color that applies to the sheet background and sheet controls.

$sel:successForeground:UIColorPalette', uIColorPalette_successForeground - The foreground color that applies to any text or other elements that appear over the success color.

$sel:dimension:UIColorPalette', uIColorPalette_dimension - The color that applies to the names of fields that are identified as dimensions.

$sel:dimensionForeground:UIColorPalette', uIColorPalette_dimensionForeground - The foreground color that applies to any text or other elements that appear over the dimension color.

$sel:success:UIColorPalette', uIColorPalette_success - The color that applies to success messages, for example the check mark for a successful download.

$sel:accent:UIColorPalette', uIColorPalette_accent - This color is that applies to selected states and buttons.

$sel:secondaryForeground:UIColorPalette', uIColorPalette_secondaryForeground - The foreground color that applies to any sheet title, sheet control text, or UI that appears over the secondary background.

$sel:primaryForeground:UIColorPalette', uIColorPalette_primaryForeground - The color of text and other foreground elements that appear over the primary background regions, such as grid lines, borders, table banding, icons, and so on.

$sel:warning:UIColorPalette', uIColorPalette_warning - This color that applies to warning and informational messages.

$sel:primaryBackground:UIColorPalette', uIColorPalette_primaryBackground - The background color that applies to visuals and other high emphasis UI.

$sel:accentForeground:UIColorPalette', uIColorPalette_accentForeground - The foreground color that applies to any text or other elements that appear over the accent color.

$sel:measure:UIColorPalette', uIColorPalette_measure - The color that applies to the names of fields that are identified as measures.

$sel:dangerForeground:UIColorPalette', uIColorPalette_dangerForeground - The foreground color that applies to any text or other elements that appear over the error color.

$sel:measureForeground:UIColorPalette', uIColorPalette_measureForeground - The foreground color that applies to any text or other elements that appear over the measure color.

$sel:danger:UIColorPalette', uIColorPalette_danger - The color that applies to error messages.

$sel:warningForeground:UIColorPalette', uIColorPalette_warningForeground - The foreground color that applies to any text or other elements that appear over the warning color.

uIColorPalette_secondaryBackground :: Lens' UIColorPalette (Maybe Text)

The background color that applies to the sheet background and sheet controls.

uIColorPalette_successForeground :: Lens' UIColorPalette (Maybe Text)

The foreground color that applies to any text or other elements that appear over the success color.

uIColorPalette_dimension :: Lens' UIColorPalette (Maybe Text)

The color that applies to the names of fields that are identified as dimensions.

uIColorPalette_dimensionForeground :: Lens' UIColorPalette (Maybe Text)

The foreground color that applies to any text or other elements that appear over the dimension color.

uIColorPalette_success :: Lens' UIColorPalette (Maybe Text)

The color that applies to success messages, for example the check mark for a successful download.

uIColorPalette_accent :: Lens' UIColorPalette (Maybe Text)

This color is that applies to selected states and buttons.

uIColorPalette_secondaryForeground :: Lens' UIColorPalette (Maybe Text)

The foreground color that applies to any sheet title, sheet control text, or UI that appears over the secondary background.

uIColorPalette_primaryForeground :: Lens' UIColorPalette (Maybe Text)

The color of text and other foreground elements that appear over the primary background regions, such as grid lines, borders, table banding, icons, and so on.

uIColorPalette_warning :: Lens' UIColorPalette (Maybe Text)

This color that applies to warning and informational messages.

uIColorPalette_primaryBackground :: Lens' UIColorPalette (Maybe Text)

The background color that applies to visuals and other high emphasis UI.

uIColorPalette_accentForeground :: Lens' UIColorPalette (Maybe Text)

The foreground color that applies to any text or other elements that appear over the accent color.

uIColorPalette_measure :: Lens' UIColorPalette (Maybe Text)

The color that applies to the names of fields that are identified as measures.

uIColorPalette_dangerForeground :: Lens' UIColorPalette (Maybe Text)

The foreground color that applies to any text or other elements that appear over the error color.

uIColorPalette_measureForeground :: Lens' UIColorPalette (Maybe Text)

The foreground color that applies to any text or other elements that appear over the measure color.

uIColorPalette_danger :: Lens' UIColorPalette (Maybe Text)

The color that applies to error messages.

uIColorPalette_warningForeground :: Lens' UIColorPalette (Maybe Text)

The foreground color that applies to any text or other elements that appear over the warning color.