Amazonka.QuickSight.UpdateAccountCustomization

Description

Updates Amazon QuickSight customizations the current Amazon Web Services Region. Currently, the only customization you can use is a theme.

You can use customizations for your Amazon Web Services account or, if you specify a namespace, for a Amazon QuickSight namespace instead. Customizations that apply to a namespace override customizations that apply to an Amazon Web Services account. To find out which customizations apply, use the DescribeAccountCustomization API operation.

Creating a Request

data UpdateAccountCustomization

See: newUpdateAccountCustomization smart constructor.

Constructors

UpdateAccountCustomization'

Fields

namespace :: Maybe Text

The namespace that you want to update Amazon QuickSight customizations for.

awsAccountId :: Text

The ID for the Amazon Web Services account that you want to update Amazon QuickSight customizations for.

accountCustomization :: AccountCustomization

The Amazon QuickSight customizations you're updating in the current Amazon Web Services Region.

newUpdateAccountCustomization

Arguments

:: Text
$sel:awsAccountId:UpdateAccountCustomization'

-> AccountCustomization
$sel:accountCustomization:UpdateAccountCustomization'

-> UpdateAccountCustomization

Create a value of UpdateAccountCustomization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namespace:UpdateAccountCustomization', updateAccountCustomization_namespace - The namespace that you want to update Amazon QuickSight customizations for.

$sel:awsAccountId:UpdateAccountCustomization', updateAccountCustomization_awsAccountId - The ID for the Amazon Web Services account that you want to update Amazon QuickSight customizations for.

$sel:accountCustomization:UpdateAccountCustomization', updateAccountCustomization_accountCustomization - The Amazon QuickSight customizations you're updating in the current Amazon Web Services Region.

Request Lenses

updateAccountCustomization_namespace :: Lens' UpdateAccountCustomization (Maybe Text)

The namespace that you want to update Amazon QuickSight customizations for.

updateAccountCustomization_awsAccountId :: Lens' UpdateAccountCustomization Text

The ID for the Amazon Web Services account that you want to update Amazon QuickSight customizations for.

updateAccountCustomization_accountCustomization :: Lens' UpdateAccountCustomization AccountCustomization

The Amazon QuickSight customizations you're updating in the current Amazon Web Services Region.

Destructuring the Response

data UpdateAccountCustomizationResponse

See: newUpdateAccountCustomizationResponse smart constructor.

Constructors

UpdateAccountCustomizationResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

accountCustomization :: Maybe AccountCustomization

The Amazon QuickSight customizations you're updating in the current Amazon Web Services Region.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the updated customization for this Amazon Web Services account.

namespace :: Maybe Text

The namespace associated with the customization that you're updating.

awsAccountId :: Maybe Text

The ID for the Amazon Web Services account that you want to update Amazon QuickSight customizations for.

status :: Int

The HTTP status of the request.

newUpdateAccountCustomizationResponse

Arguments

:: Int
$sel:status:UpdateAccountCustomizationResponse'

-> UpdateAccountCustomizationResponse

Create a value of UpdateAccountCustomizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateAccountCustomizationResponse', updateAccountCustomizationResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:accountCustomization:UpdateAccountCustomization', updateAccountCustomizationResponse_accountCustomization - The Amazon QuickSight customizations you're updating in the current Amazon Web Services Region.

$sel:arn:UpdateAccountCustomizationResponse', updateAccountCustomizationResponse_arn - The Amazon Resource Name (ARN) for the updated customization for this Amazon Web Services account.

$sel:namespace:UpdateAccountCustomization', updateAccountCustomizationResponse_namespace - The namespace associated with the customization that you're updating.

$sel:awsAccountId:UpdateAccountCustomization', updateAccountCustomizationResponse_awsAccountId - The ID for the Amazon Web Services account that you want to update Amazon QuickSight customizations for.

$sel:status:UpdateAccountCustomizationResponse', updateAccountCustomizationResponse_status - The HTTP status of the request.

Response Lenses

updateAccountCustomizationResponse_requestId :: Lens' UpdateAccountCustomizationResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

updateAccountCustomizationResponse_accountCustomization :: Lens' UpdateAccountCustomizationResponse (Maybe AccountCustomization)

The Amazon QuickSight customizations you're updating in the current Amazon Web Services Region.

updateAccountCustomizationResponse_arn :: Lens' UpdateAccountCustomizationResponse (Maybe Text)

The Amazon Resource Name (ARN) for the updated customization for this Amazon Web Services account.

updateAccountCustomizationResponse_namespace :: Lens' UpdateAccountCustomizationResponse (Maybe Text)

The namespace associated with the customization that you're updating.

updateAccountCustomizationResponse_awsAccountId :: Lens' UpdateAccountCustomizationResponse (Maybe Text)

The ID for the Amazon Web Services account that you want to update Amazon QuickSight customizations for.

updateAccountCustomizationResponse_status :: Lens' UpdateAccountCustomizationResponse Int

The HTTP status of the request.