Amazonka.QuickSight.UpdateDashboard

Description

Updates a dashboard in an Amazon Web Services account.

Updating a Dashboard creates a new dashboard version but does not immediately publish the new version. You can update the published version of a dashboard by using the UpdateDashboardPublishedVersion API operation.

Creating a Request

data UpdateDashboard

See: newUpdateDashboard smart constructor.

Constructors

UpdateDashboard'

Fields

themeArn :: Maybe Text

The Amazon Resource Name (ARN) of the theme that is being used for this dashboard. If you add a value for this field, it overrides the value that was originally associated with the entity. The theme ARN must exist in the same Amazon Web Services account where you create the dashboard.

dashboardPublishOptions :: Maybe DashboardPublishOptions

Options for publishing the dashboard when you create it:

AvailabilityStatus for AdHocFilteringOption - This status can be either ENABLED or DISABLED. When this is set to DISABLED, Amazon QuickSight disables the left filter pane on the published dashboard, which can be used for ad hoc (one-time) filtering. This option is ENABLED by default.

AvailabilityStatus for ExportToCSVOption - This status can be either ENABLED or DISABLED. The visual option to export data to .CSV format isn't enabled when this is set to DISABLED. This option is ENABLED by default.

VisibilityState for SheetControlsOption - This visibility state can be either COLLAPSED or EXPANDED. This option is COLLAPSED by default.

versionDescription :: Maybe Text

A description for the first version of the dashboard being created.

parameters :: Maybe Parameters

A structure that contains the parameters of the dashboard. These are parameter overrides for a dashboard. A dashboard can have any type of parameters, and some parameters might accept multiple values.

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the dashboard that you're updating.

dashboardId :: Text

The ID for the dashboard.

name :: Text

The display name of the dashboard.

sourceEntity :: DashboardSourceEntity

The entity that you are using as a source when you update the dashboard. In SourceEntity, you specify the type of object you're using as source. You can only update a dashboard from a template, so you use a SourceTemplate entity. If you need to update a dashboard from an analysis, first convert the analysis to a template by using the CreateTemplate API operation. For SourceTemplate, specify the Amazon Resource Name (ARN) of the source template. The SourceTemplate ARN can contain any Amazon Web Services account and any Amazon QuickSight-supported Amazon Web Services Region.

Use the DataSetReferences entity within SourceTemplate to list the replacement datasets for the placeholders listed in the original. The schema in each dataset must match its placeholder.

newUpdateDashboard

Arguments

:: Text
$sel:awsAccountId:UpdateDashboard'

-> Text
$sel:dashboardId:UpdateDashboard'

-> Text
$sel:name:UpdateDashboard'

-> DashboardSourceEntity
$sel:sourceEntity:UpdateDashboard'

-> UpdateDashboard

Create a value of UpdateDashboard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:themeArn:UpdateDashboard', updateDashboard_themeArn - The Amazon Resource Name (ARN) of the theme that is being used for this dashboard. If you add a value for this field, it overrides the value that was originally associated with the entity. The theme ARN must exist in the same Amazon Web Services account where you create the dashboard.

$sel:dashboardPublishOptions:UpdateDashboard', updateDashboard_dashboardPublishOptions - Options for publishing the dashboard when you create it:

AvailabilityStatus for AdHocFilteringOption - This status can be either ENABLED or DISABLED. When this is set to DISABLED, Amazon QuickSight disables the left filter pane on the published dashboard, which can be used for ad hoc (one-time) filtering. This option is ENABLED by default.

AvailabilityStatus for ExportToCSVOption - This status can be either ENABLED or DISABLED. The visual option to export data to .CSV format isn't enabled when this is set to DISABLED. This option is ENABLED by default.

VisibilityState for SheetControlsOption - This visibility state can be either COLLAPSED or EXPANDED. This option is COLLAPSED by default.

$sel:versionDescription:UpdateDashboard', updateDashboard_versionDescription - A description for the first version of the dashboard being created.

$sel:parameters:UpdateDashboard', updateDashboard_parameters - A structure that contains the parameters of the dashboard. These are parameter overrides for a dashboard. A dashboard can have any type of parameters, and some parameters might accept multiple values.

$sel:awsAccountId:UpdateDashboard', updateDashboard_awsAccountId - The ID of the Amazon Web Services account that contains the dashboard that you're updating.

$sel:dashboardId:UpdateDashboard', updateDashboard_dashboardId - The ID for the dashboard.

$sel:name:UpdateDashboard', updateDashboard_name - The display name of the dashboard.

$sel:sourceEntity:UpdateDashboard', updateDashboard_sourceEntity - The entity that you are using as a source when you update the dashboard. In SourceEntity, you specify the type of object you're using as source. You can only update a dashboard from a template, so you use a SourceTemplate entity. If you need to update a dashboard from an analysis, first convert the analysis to a template by using the CreateTemplate API operation. For SourceTemplate, specify the Amazon Resource Name (ARN) of the source template. The SourceTemplate ARN can contain any Amazon Web Services account and any Amazon QuickSight-supported Amazon Web Services Region.

Use the DataSetReferences entity within SourceTemplate to list the replacement datasets for the placeholders listed in the original. The schema in each dataset must match its placeholder.

Request Lenses

updateDashboard_themeArn :: Lens' UpdateDashboard (Maybe Text)

The Amazon Resource Name (ARN) of the theme that is being used for this dashboard. If you add a value for this field, it overrides the value that was originally associated with the entity. The theme ARN must exist in the same Amazon Web Services account where you create the dashboard.

updateDashboard_dashboardPublishOptions :: Lens' UpdateDashboard (Maybe DashboardPublishOptions)

Options for publishing the dashboard when you create it:

AvailabilityStatus for AdHocFilteringOption - This status can be either ENABLED or DISABLED. When this is set to DISABLED, Amazon QuickSight disables the left filter pane on the published dashboard, which can be used for ad hoc (one-time) filtering. This option is ENABLED by default.

AvailabilityStatus for ExportToCSVOption - This status can be either ENABLED or DISABLED. The visual option to export data to .CSV format isn't enabled when this is set to DISABLED. This option is ENABLED by default.

VisibilityState for SheetControlsOption - This visibility state can be either COLLAPSED or EXPANDED. This option is COLLAPSED by default.

updateDashboard_versionDescription :: Lens' UpdateDashboard (Maybe Text)

A description for the first version of the dashboard being created.

updateDashboard_parameters :: Lens' UpdateDashboard (Maybe Parameters)

A structure that contains the parameters of the dashboard. These are parameter overrides for a dashboard. A dashboard can have any type of parameters, and some parameters might accept multiple values.

updateDashboard_awsAccountId :: Lens' UpdateDashboard Text

The ID of the Amazon Web Services account that contains the dashboard that you're updating.

updateDashboard_dashboardId :: Lens' UpdateDashboard Text

The ID for the dashboard.

updateDashboard_name :: Lens' UpdateDashboard Text

The display name of the dashboard.

updateDashboard_sourceEntity :: Lens' UpdateDashboard DashboardSourceEntity

The entity that you are using as a source when you update the dashboard. In SourceEntity, you specify the type of object you're using as source. You can only update a dashboard from a template, so you use a SourceTemplate entity. If you need to update a dashboard from an analysis, first convert the analysis to a template by using the CreateTemplate API operation. For SourceTemplate, specify the Amazon Resource Name (ARN) of the source template. The SourceTemplate ARN can contain any Amazon Web Services account and any Amazon QuickSight-supported Amazon Web Services Region.

Use the DataSetReferences entity within SourceTemplate to list the replacement datasets for the placeholders listed in the original. The schema in each dataset must match its placeholder.

Destructuring the Response

data UpdateDashboardResponse

See: newUpdateDashboardResponse smart constructor.

Constructors

UpdateDashboardResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

status :: Maybe Int

The HTTP status of the request.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the resource.

creationStatus :: Maybe ResourceStatus

The creation status of the request.

dashboardId :: Maybe Text

The ID for the dashboard.

versionArn :: Maybe Text

The ARN of the dashboard, including the version number.

httpStatus :: Int

The response's http status code.

newUpdateDashboardResponse

Arguments

:: Int
$sel:httpStatus:UpdateDashboardResponse'

-> UpdateDashboardResponse

Create a value of UpdateDashboardResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateDashboardResponse', updateDashboardResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:status:UpdateDashboardResponse', updateDashboardResponse_status - The HTTP status of the request.

$sel:arn:UpdateDashboardResponse', updateDashboardResponse_arn - The Amazon Resource Name (ARN) of the resource.

$sel:creationStatus:UpdateDashboardResponse', updateDashboardResponse_creationStatus - The creation status of the request.

$sel:dashboardId:UpdateDashboard', updateDashboardResponse_dashboardId - The ID for the dashboard.

$sel:versionArn:UpdateDashboardResponse', updateDashboardResponse_versionArn - The ARN of the dashboard, including the version number.

$sel:httpStatus:UpdateDashboardResponse', updateDashboardResponse_httpStatus - The response's http status code.

Response Lenses

updateDashboardResponse_requestId :: Lens' UpdateDashboardResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

updateDashboardResponse_status :: Lens' UpdateDashboardResponse (Maybe Int)

The HTTP status of the request.

updateDashboardResponse_arn :: Lens' UpdateDashboardResponse (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

updateDashboardResponse_creationStatus :: Lens' UpdateDashboardResponse (Maybe ResourceStatus)

The creation status of the request.

updateDashboardResponse_dashboardId :: Lens' UpdateDashboardResponse (Maybe Text)

The ID for the dashboard.

updateDashboardResponse_versionArn :: Lens' UpdateDashboardResponse (Maybe Text)

The ARN of the dashboard, including the version number.

updateDashboardResponse_httpStatus :: Lens' UpdateDashboardResponse Int

The response's http status code.