Amazonka.QuickSight.UpdateDashboardPermissions

Description

Updates read and write permissions on a dashboard.

Creating a Request

data UpdateDashboardPermissions

See: newUpdateDashboardPermissions smart constructor.

Constructors

UpdateDashboardPermissions'

Fields

revokePermissions :: Maybe [ResourcePermission]

The permissions that you want to revoke from this resource.

grantPermissions :: Maybe [ResourcePermission]

The permissions that you want to grant on this resource.

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the dashboard whose permissions you're updating.

dashboardId :: Text

The ID for the dashboard.

newUpdateDashboardPermissions

Arguments

:: Text
$sel:awsAccountId:UpdateDashboardPermissions'

-> Text
$sel:dashboardId:UpdateDashboardPermissions'

-> UpdateDashboardPermissions

Create a value of UpdateDashboardPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revokePermissions:UpdateDashboardPermissions', updateDashboardPermissions_revokePermissions - The permissions that you want to revoke from this resource.

$sel:grantPermissions:UpdateDashboardPermissions', updateDashboardPermissions_grantPermissions - The permissions that you want to grant on this resource.

$sel:awsAccountId:UpdateDashboardPermissions', updateDashboardPermissions_awsAccountId - The ID of the Amazon Web Services account that contains the dashboard whose permissions you're updating.

$sel:dashboardId:UpdateDashboardPermissions', updateDashboardPermissions_dashboardId - The ID for the dashboard.

Request Lenses

updateDashboardPermissions_revokePermissions :: Lens' UpdateDashboardPermissions (Maybe [ResourcePermission])

The permissions that you want to revoke from this resource.

updateDashboardPermissions_grantPermissions :: Lens' UpdateDashboardPermissions (Maybe [ResourcePermission])

The permissions that you want to grant on this resource.

updateDashboardPermissions_awsAccountId :: Lens' UpdateDashboardPermissions Text

The ID of the Amazon Web Services account that contains the dashboard whose permissions you're updating.

updateDashboardPermissions_dashboardId :: Lens' UpdateDashboardPermissions Text

The ID for the dashboard.

Destructuring the Response

data UpdateDashboardPermissionsResponse

See: newUpdateDashboardPermissionsResponse smart constructor.

Constructors

UpdateDashboardPermissionsResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

dashboardId :: Maybe Text

The ID for the dashboard.

permissions :: Maybe (NonEmpty ResourcePermission)

Information about the permissions on the dashboard.

dashboardArn :: Maybe Text

The Amazon Resource Name (ARN) of the dashboard.

status :: Int

The HTTP status of the request.

newUpdateDashboardPermissionsResponse

Arguments

:: Int
$sel:status:UpdateDashboardPermissionsResponse'

-> UpdateDashboardPermissionsResponse

Create a value of UpdateDashboardPermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateDashboardPermissionsResponse', updateDashboardPermissionsResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:dashboardId:UpdateDashboardPermissions', updateDashboardPermissionsResponse_dashboardId - The ID for the dashboard.

$sel:permissions:UpdateDashboardPermissionsResponse', updateDashboardPermissionsResponse_permissions - Information about the permissions on the dashboard.

$sel:dashboardArn:UpdateDashboardPermissionsResponse', updateDashboardPermissionsResponse_dashboardArn - The Amazon Resource Name (ARN) of the dashboard.

$sel:status:UpdateDashboardPermissionsResponse', updateDashboardPermissionsResponse_status - The HTTP status of the request.

Response Lenses

updateDashboardPermissionsResponse_requestId :: Lens' UpdateDashboardPermissionsResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

updateDashboardPermissionsResponse_dashboardId :: Lens' UpdateDashboardPermissionsResponse (Maybe Text)

The ID for the dashboard.

updateDashboardPermissionsResponse_permissions :: Lens' UpdateDashboardPermissionsResponse (Maybe (NonEmpty ResourcePermission))

Information about the permissions on the dashboard.

updateDashboardPermissionsResponse_dashboardArn :: Lens' UpdateDashboardPermissionsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the dashboard.

updateDashboardPermissionsResponse_status :: Lens' UpdateDashboardPermissionsResponse Int

The HTTP status of the request.