Amazonka.QuickSight.UpdateDataSet

Description

Updates a dataset.

Creating a Request

data UpdateDataSet

See: newUpdateDataSet smart constructor.

Constructors

UpdateDataSet'

Fields

fieldFolders :: Maybe (HashMap Text FieldFolder)

The folder that contains fields and nested subfolders for your dataset.

columnGroups :: Maybe (NonEmpty ColumnGroup)

Groupings of columns that work together in certain Amazon QuickSight features. Currently, only geospatial hierarchy is supported.

dataSetUsageConfiguration :: Maybe DataSetUsageConfiguration

rowLevelPermissionTagConfiguration :: Maybe RowLevelPermissionTagConfiguration

The configuration of tags on a dataset to set row-level security. Row-level security tags are currently supported for anonymous embedding only.

rowLevelPermissionDataSet :: Maybe RowLevelPermissionDataSet

The row-level security configuration for the data you want to create.

columnLevelPermissionRules :: Maybe (NonEmpty ColumnLevelPermissionRule)

A set of one or more definitions of a ColumnLevelPermissionRule .

logicalTableMap :: Maybe (HashMap Text LogicalTable)

Configures the combination and transformation of the data from the physical tables.

awsAccountId :: Text

The Amazon Web Services account ID.

dataSetId :: Text

The ID for the dataset that you want to update. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

name :: Text

The display name for the dataset.

physicalTableMap :: HashMap Text PhysicalTable

Declares the physical tables that are available in the underlying data sources.

importMode :: DataSetImportMode

Indicates whether you want to import the data into SPICE.

newUpdateDataSet

Arguments

:: Text
$sel:awsAccountId:UpdateDataSet'

-> Text
$sel:dataSetId:UpdateDataSet'

-> Text
$sel:name:UpdateDataSet'

-> DataSetImportMode
$sel:importMode:UpdateDataSet'

-> UpdateDataSet

Create a value of UpdateDataSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fieldFolders:UpdateDataSet', updateDataSet_fieldFolders - The folder that contains fields and nested subfolders for your dataset.

$sel:columnGroups:UpdateDataSet', updateDataSet_columnGroups - Groupings of columns that work together in certain Amazon QuickSight features. Currently, only geospatial hierarchy is supported.

$sel:dataSetUsageConfiguration:UpdateDataSet', updateDataSet_dataSetUsageConfiguration - Undocumented member.

$sel:rowLevelPermissionTagConfiguration:UpdateDataSet', updateDataSet_rowLevelPermissionTagConfiguration - The configuration of tags on a dataset to set row-level security. Row-level security tags are currently supported for anonymous embedding only.

$sel:rowLevelPermissionDataSet:UpdateDataSet', updateDataSet_rowLevelPermissionDataSet - The row-level security configuration for the data you want to create.

$sel:columnLevelPermissionRules:UpdateDataSet', updateDataSet_columnLevelPermissionRules - A set of one or more definitions of a ColumnLevelPermissionRule .

$sel:logicalTableMap:UpdateDataSet', updateDataSet_logicalTableMap - Configures the combination and transformation of the data from the physical tables.

$sel:awsAccountId:UpdateDataSet', updateDataSet_awsAccountId - The Amazon Web Services account ID.

$sel:dataSetId:UpdateDataSet', updateDataSet_dataSetId - The ID for the dataset that you want to update. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

$sel:name:UpdateDataSet', updateDataSet_name - The display name for the dataset.

$sel:physicalTableMap:UpdateDataSet', updateDataSet_physicalTableMap - Declares the physical tables that are available in the underlying data sources.

$sel:importMode:UpdateDataSet', updateDataSet_importMode - Indicates whether you want to import the data into SPICE.

Request Lenses

updateDataSet_fieldFolders :: Lens' UpdateDataSet (Maybe (HashMap Text FieldFolder))

The folder that contains fields and nested subfolders for your dataset.

updateDataSet_columnGroups :: Lens' UpdateDataSet (Maybe (NonEmpty ColumnGroup))

Groupings of columns that work together in certain Amazon QuickSight features. Currently, only geospatial hierarchy is supported.

updateDataSet_dataSetUsageConfiguration :: Lens' UpdateDataSet (Maybe DataSetUsageConfiguration)

Undocumented member.

updateDataSet_rowLevelPermissionTagConfiguration :: Lens' UpdateDataSet (Maybe RowLevelPermissionTagConfiguration)

The configuration of tags on a dataset to set row-level security. Row-level security tags are currently supported for anonymous embedding only.

updateDataSet_rowLevelPermissionDataSet :: Lens' UpdateDataSet (Maybe RowLevelPermissionDataSet)

The row-level security configuration for the data you want to create.

updateDataSet_columnLevelPermissionRules :: Lens' UpdateDataSet (Maybe (NonEmpty ColumnLevelPermissionRule))

A set of one or more definitions of a ColumnLevelPermissionRule .

updateDataSet_logicalTableMap :: Lens' UpdateDataSet (Maybe (HashMap Text LogicalTable))

Configures the combination and transformation of the data from the physical tables.

updateDataSet_awsAccountId :: Lens' UpdateDataSet Text

The Amazon Web Services account ID.

updateDataSet_dataSetId :: Lens' UpdateDataSet Text

The ID for the dataset that you want to update. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

updateDataSet_name :: Lens' UpdateDataSet Text

The display name for the dataset.

updateDataSet_physicalTableMap :: Lens' UpdateDataSet (HashMap Text PhysicalTable)

Declares the physical tables that are available in the underlying data sources.

updateDataSet_importMode :: Lens' UpdateDataSet DataSetImportMode

Indicates whether you want to import the data into SPICE.

Destructuring the Response

data UpdateDataSetResponse

See: newUpdateDataSetResponse smart constructor.

Constructors

UpdateDataSetResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

ingestionArn :: Maybe Text

The ARN for the ingestion, which is triggered as a result of dataset creation if the import mode is SPICE.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset.

ingestionId :: Maybe Text

The ID of the ingestion, which is triggered as a result of dataset creation if the import mode is SPICE.

dataSetId :: Maybe Text

The ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

status :: Int

The HTTP status of the request.

newUpdateDataSetResponse

Arguments

:: Int
$sel:status:UpdateDataSetResponse'

-> UpdateDataSetResponse

Create a value of UpdateDataSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateDataSetResponse', updateDataSetResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:ingestionArn:UpdateDataSetResponse', updateDataSetResponse_ingestionArn - The ARN for the ingestion, which is triggered as a result of dataset creation if the import mode is SPICE.

$sel:arn:UpdateDataSetResponse', updateDataSetResponse_arn - The Amazon Resource Name (ARN) of the dataset.

$sel:ingestionId:UpdateDataSetResponse', updateDataSetResponse_ingestionId - The ID of the ingestion, which is triggered as a result of dataset creation if the import mode is SPICE.

$sel:dataSetId:UpdateDataSet', updateDataSetResponse_dataSetId - The ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

$sel:status:UpdateDataSetResponse', updateDataSetResponse_status - The HTTP status of the request.

Response Lenses

updateDataSetResponse_requestId :: Lens' UpdateDataSetResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

updateDataSetResponse_ingestionArn :: Lens' UpdateDataSetResponse (Maybe Text)

The ARN for the ingestion, which is triggered as a result of dataset creation if the import mode is SPICE.

updateDataSetResponse_arn :: Lens' UpdateDataSetResponse (Maybe Text)

The Amazon Resource Name (ARN) of the dataset.

updateDataSetResponse_ingestionId :: Lens' UpdateDataSetResponse (Maybe Text)

The ID of the ingestion, which is triggered as a result of dataset creation if the import mode is SPICE.

updateDataSetResponse_dataSetId :: Lens' UpdateDataSetResponse (Maybe Text)

The ID for the dataset that you want to create. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

updateDataSetResponse_status :: Lens' UpdateDataSetResponse Int

The HTTP status of the request.