Amazonka.QuickSight.UpdateIAMPolicyAssignment

Description

Updates an existing IAMpolicy assignment. This operation updates only the optional parameter or parameters that are specified in the request. This overwrites all of the users included in Identities.

Creating a Request

data UpdateIAMPolicyAssignment

See: newUpdateIAMPolicyAssignment smart constructor.

Constructors

UpdateIAMPolicyAssignment'

Fields

assignmentStatus :: Maybe AssignmentStatus

The status of the assignment. Possible values are as follows:

ENABLED - Anything specified in this assignment is used when creating the data source.

DISABLED - This assignment isn't used when creating the data source.

DRAFT - This assignment is an unfinished draft and isn't used when creating the data source.

policyArn :: Maybe Text

The ARN for the IAMpolicy to apply to the Amazon QuickSight users and groups specified in this assignment.

identities :: Maybe (HashMap Text [Text])

The Amazon QuickSight users, groups, or both that you want to assign the policy to.

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the IAMpolicy assignment.

assignmentName :: Text

The name of the assignment, also called a rule. This name must be unique within an Amazon Web Services account.

namespace :: Text

The namespace of the assignment.

newUpdateIAMPolicyAssignment

Arguments

:: Text
$sel:awsAccountId:UpdateIAMPolicyAssignment'

-> Text
$sel:assignmentName:UpdateIAMPolicyAssignment'

-> Text
$sel:namespace:UpdateIAMPolicyAssignment'

-> UpdateIAMPolicyAssignment

Create a value of UpdateIAMPolicyAssignment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assignmentStatus:UpdateIAMPolicyAssignment', updateIAMPolicyAssignment_assignmentStatus - The status of the assignment. Possible values are as follows:

ENABLED - Anything specified in this assignment is used when creating the data source.

DISABLED - This assignment isn't used when creating the data source.

DRAFT - This assignment is an unfinished draft and isn't used when creating the data source.

$sel:policyArn:UpdateIAMPolicyAssignment', updateIAMPolicyAssignment_policyArn - The ARN for the IAMpolicy to apply to the Amazon QuickSight users and groups specified in this assignment.

$sel:identities:UpdateIAMPolicyAssignment', updateIAMPolicyAssignment_identities - The Amazon QuickSight users, groups, or both that you want to assign the policy to.

$sel:awsAccountId:UpdateIAMPolicyAssignment', updateIAMPolicyAssignment_awsAccountId - The ID of the Amazon Web Services account that contains the IAMpolicy assignment.

$sel:assignmentName:UpdateIAMPolicyAssignment', updateIAMPolicyAssignment_assignmentName - The name of the assignment, also called a rule. This name must be unique within an Amazon Web Services account.

$sel:namespace:UpdateIAMPolicyAssignment', updateIAMPolicyAssignment_namespace - The namespace of the assignment.

Request Lenses

updateIAMPolicyAssignment_assignmentStatus :: Lens' UpdateIAMPolicyAssignment (Maybe AssignmentStatus)

The status of the assignment. Possible values are as follows:

ENABLED - Anything specified in this assignment is used when creating the data source.

DISABLED - This assignment isn't used when creating the data source.

DRAFT - This assignment is an unfinished draft and isn't used when creating the data source.

updateIAMPolicyAssignment_policyArn :: Lens' UpdateIAMPolicyAssignment (Maybe Text)

The ARN for the IAMpolicy to apply to the Amazon QuickSight users and groups specified in this assignment.

updateIAMPolicyAssignment_identities :: Lens' UpdateIAMPolicyAssignment (Maybe (HashMap Text [Text]))

The Amazon QuickSight users, groups, or both that you want to assign the policy to.

updateIAMPolicyAssignment_awsAccountId :: Lens' UpdateIAMPolicyAssignment Text

The ID of the Amazon Web Services account that contains the IAMpolicy assignment.

updateIAMPolicyAssignment_assignmentName :: Lens' UpdateIAMPolicyAssignment Text

The name of the assignment, also called a rule. This name must be unique within an Amazon Web Services account.

updateIAMPolicyAssignment_namespace :: Lens' UpdateIAMPolicyAssignment Text

The namespace of the assignment.

Destructuring the Response

data UpdateIAMPolicyAssignmentResponse

See: newUpdateIAMPolicyAssignmentResponse smart constructor.

Constructors

UpdateIAMPolicyAssignmentResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

assignmentName :: Maybe Text

The name of the assignment or rule.

assignmentStatus :: Maybe AssignmentStatus

The status of the assignment. Possible values are as follows:

ENABLED - Anything specified in this assignment is used when creating the data source.

DISABLED - This assignment isn't used when creating the data source.

DRAFT - This assignment is an unfinished draft and isn't used when creating the data source.

policyArn :: Maybe Text

The ARN for the IAMpolicy applied to the Amazon QuickSight users and groups specified in this assignment.

identities :: Maybe (HashMap Text [Text])

The Amazon QuickSight users, groups, or both that the IAMpolicy is assigned to.

assignmentId :: Maybe Text

The ID of the assignment.

status :: Int

The HTTP status of the request.

newUpdateIAMPolicyAssignmentResponse

Arguments

:: Int
$sel:status:UpdateIAMPolicyAssignmentResponse'

-> UpdateIAMPolicyAssignmentResponse

Create a value of UpdateIAMPolicyAssignmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateIAMPolicyAssignmentResponse', updateIAMPolicyAssignmentResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:assignmentName:UpdateIAMPolicyAssignment', updateIAMPolicyAssignmentResponse_assignmentName - The name of the assignment or rule.

$sel:assignmentStatus:UpdateIAMPolicyAssignment', updateIAMPolicyAssignmentResponse_assignmentStatus - The status of the assignment. Possible values are as follows:

ENABLED - Anything specified in this assignment is used when creating the data source.

DISABLED - This assignment isn't used when creating the data source.

DRAFT - This assignment is an unfinished draft and isn't used when creating the data source.

$sel:policyArn:UpdateIAMPolicyAssignment', updateIAMPolicyAssignmentResponse_policyArn - The ARN for the IAMpolicy applied to the Amazon QuickSight users and groups specified in this assignment.

$sel:identities:UpdateIAMPolicyAssignment', updateIAMPolicyAssignmentResponse_identities - The Amazon QuickSight users, groups, or both that the IAMpolicy is assigned to.

$sel:assignmentId:UpdateIAMPolicyAssignmentResponse', updateIAMPolicyAssignmentResponse_assignmentId - The ID of the assignment.

$sel:status:UpdateIAMPolicyAssignmentResponse', updateIAMPolicyAssignmentResponse_status - The HTTP status of the request.

Response Lenses

updateIAMPolicyAssignmentResponse_requestId :: Lens' UpdateIAMPolicyAssignmentResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

updateIAMPolicyAssignmentResponse_assignmentName :: Lens' UpdateIAMPolicyAssignmentResponse (Maybe Text)

The name of the assignment or rule.

updateIAMPolicyAssignmentResponse_assignmentStatus :: Lens' UpdateIAMPolicyAssignmentResponse (Maybe AssignmentStatus)

The status of the assignment. Possible values are as follows:

ENABLED - Anything specified in this assignment is used when creating the data source.

DISABLED - This assignment isn't used when creating the data source.

DRAFT - This assignment is an unfinished draft and isn't used when creating the data source.

updateIAMPolicyAssignmentResponse_policyArn :: Lens' UpdateIAMPolicyAssignmentResponse (Maybe Text)

The ARN for the IAMpolicy applied to the Amazon QuickSight users and groups specified in this assignment.

updateIAMPolicyAssignmentResponse_identities :: Lens' UpdateIAMPolicyAssignmentResponse (Maybe (HashMap Text [Text]))

The Amazon QuickSight users, groups, or both that the IAMpolicy is assigned to.

updateIAMPolicyAssignmentResponse_assignmentId :: Lens' UpdateIAMPolicyAssignmentResponse (Maybe Text)

The ID of the assignment.

updateIAMPolicyAssignmentResponse_status :: Lens' UpdateIAMPolicyAssignmentResponse Int

The HTTP status of the request.