Amazonka.QuickSight.UpdateTemplateAlias

Description

Updates the template alias of a template.

Creating a Request

data UpdateTemplateAlias

See: newUpdateTemplateAlias smart constructor.

Constructors

UpdateTemplateAlias'

Fields

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the template alias that you're updating.

templateId :: Text

The ID for the template.

aliasName :: Text

The alias of the template that you want to update. If you name a specific alias, you update the version that the alias points to. You can specify the latest version of the template by providing the keyword $LATEST in the AliasName parameter. The keyword $PUBLISHED doesn't apply to templates.

templateVersionNumber :: Natural

The version number of the template.

newUpdateTemplateAlias

Arguments

:: Text
$sel:awsAccountId:UpdateTemplateAlias'

-> Text
$sel:templateId:UpdateTemplateAlias'

-> Text
$sel:aliasName:UpdateTemplateAlias'

-> Natural
$sel:templateVersionNumber:UpdateTemplateAlias'

-> UpdateTemplateAlias

Create a value of UpdateTemplateAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:UpdateTemplateAlias', updateTemplateAlias_awsAccountId - The ID of the Amazon Web Services account that contains the template alias that you're updating.

$sel:templateId:UpdateTemplateAlias', updateTemplateAlias_templateId - The ID for the template.

$sel:aliasName:UpdateTemplateAlias', updateTemplateAlias_aliasName - The alias of the template that you want to update. If you name a specific alias, you update the version that the alias points to. You can specify the latest version of the template by providing the keyword $LATEST in the AliasName parameter. The keyword $PUBLISHED doesn't apply to templates.

$sel:templateVersionNumber:UpdateTemplateAlias', updateTemplateAlias_templateVersionNumber - The version number of the template.

Request Lenses

updateTemplateAlias_awsAccountId :: Lens' UpdateTemplateAlias Text

The ID of the Amazon Web Services account that contains the template alias that you're updating.

updateTemplateAlias_templateId :: Lens' UpdateTemplateAlias Text

The ID for the template.

updateTemplateAlias_aliasName :: Lens' UpdateTemplateAlias Text

The alias of the template that you want to update. If you name a specific alias, you update the version that the alias points to. You can specify the latest version of the template by providing the keyword $LATEST in the AliasName parameter. The keyword $PUBLISHED doesn't apply to templates.

updateTemplateAlias_templateVersionNumber :: Lens' UpdateTemplateAlias Natural

The version number of the template.

Destructuring the Response

data UpdateTemplateAliasResponse

See: newUpdateTemplateAliasResponse smart constructor.

Constructors

UpdateTemplateAliasResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

templateAlias :: Maybe TemplateAlias

The template alias.

status :: Int

The HTTP status of the request.

newUpdateTemplateAliasResponse

Arguments

:: Int
$sel:status:UpdateTemplateAliasResponse'

-> UpdateTemplateAliasResponse

Create a value of UpdateTemplateAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateTemplateAliasResponse', updateTemplateAliasResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:templateAlias:UpdateTemplateAliasResponse', updateTemplateAliasResponse_templateAlias - The template alias.

$sel:status:UpdateTemplateAliasResponse', updateTemplateAliasResponse_status - The HTTP status of the request.

Response Lenses

updateTemplateAliasResponse_requestId :: Lens' UpdateTemplateAliasResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

updateTemplateAliasResponse_templateAlias :: Lens' UpdateTemplateAliasResponse (Maybe TemplateAlias)

The template alias.

updateTemplateAliasResponse_status :: Lens' UpdateTemplateAliasResponse Int

The HTTP status of the request.