Amazonka.QuickSight.UpdateTheme

Description

Updates a theme.

Creating a Request

data UpdateTheme

See: newUpdateTheme smart constructor.

Constructors

UpdateTheme'

Fields

versionDescription :: Maybe Text

A description of the theme version that you're updating Every time that you call UpdateTheme, you create a new version of the theme. Each version of the theme maintains a description of the version in VersionDescription.

name :: Maybe Text

The name for the theme.

configuration :: Maybe ThemeConfiguration

The theme configuration, which contains the theme display properties.

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the theme that you're updating.

themeId :: Text

The ID for the theme.

baseThemeId :: Text

The theme ID, defined by Amazon QuickSight, that a custom theme inherits from. All themes initially inherit from a default Amazon QuickSight theme.

newUpdateTheme

Arguments

:: Text
$sel:awsAccountId:UpdateTheme'

-> Text
$sel:themeId:UpdateTheme'

-> Text
$sel:baseThemeId:UpdateTheme'

-> UpdateTheme

Create a value of UpdateTheme with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionDescription:UpdateTheme', updateTheme_versionDescription - A description of the theme version that you're updating Every time that you call UpdateTheme, you create a new version of the theme. Each version of the theme maintains a description of the version in VersionDescription.

$sel:name:UpdateTheme', updateTheme_name - The name for the theme.

$sel:configuration:UpdateTheme', updateTheme_configuration - The theme configuration, which contains the theme display properties.

$sel:awsAccountId:UpdateTheme', updateTheme_awsAccountId - The ID of the Amazon Web Services account that contains the theme that you're updating.

$sel:themeId:UpdateTheme', updateTheme_themeId - The ID for the theme.

$sel:baseThemeId:UpdateTheme', updateTheme_baseThemeId - The theme ID, defined by Amazon QuickSight, that a custom theme inherits from. All themes initially inherit from a default Amazon QuickSight theme.

Request Lenses

updateTheme_versionDescription :: Lens' UpdateTheme (Maybe Text)

A description of the theme version that you're updating Every time that you call UpdateTheme, you create a new version of the theme. Each version of the theme maintains a description of the version in VersionDescription.

updateTheme_name :: Lens' UpdateTheme (Maybe Text)

The name for the theme.

updateTheme_configuration :: Lens' UpdateTheme (Maybe ThemeConfiguration)

The theme configuration, which contains the theme display properties.

updateTheme_awsAccountId :: Lens' UpdateTheme Text

The ID of the Amazon Web Services account that contains the theme that you're updating.

updateTheme_themeId :: Lens' UpdateTheme Text

The ID for the theme.

updateTheme_baseThemeId :: Lens' UpdateTheme Text

The theme ID, defined by Amazon QuickSight, that a custom theme inherits from. All themes initially inherit from a default Amazon QuickSight theme.

Destructuring the Response

data UpdateThemeResponse

See: newUpdateThemeResponse smart constructor.

Constructors

UpdateThemeResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

themeId :: Maybe Text

The ID for the theme.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the theme.

creationStatus :: Maybe ResourceStatus

The creation status of the theme.

versionArn :: Maybe Text

The Amazon Resource Name (ARN) for the new version of the theme.

status :: Int

The HTTP status of the request.

newUpdateThemeResponse

Arguments

:: Int
$sel:status:UpdateThemeResponse'

-> UpdateThemeResponse

Create a value of UpdateThemeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateThemeResponse', updateThemeResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:themeId:UpdateTheme', updateThemeResponse_themeId - The ID for the theme.

$sel:arn:UpdateThemeResponse', updateThemeResponse_arn - The Amazon Resource Name (ARN) for the theme.

$sel:creationStatus:UpdateThemeResponse', updateThemeResponse_creationStatus - The creation status of the theme.

$sel:versionArn:UpdateThemeResponse', updateThemeResponse_versionArn - The Amazon Resource Name (ARN) for the new version of the theme.

$sel:status:UpdateThemeResponse', updateThemeResponse_status - The HTTP status of the request.

Response Lenses

updateThemeResponse_requestId :: Lens' UpdateThemeResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

updateThemeResponse_themeId :: Lens' UpdateThemeResponse (Maybe Text)

The ID for the theme.

updateThemeResponse_arn :: Lens' UpdateThemeResponse (Maybe Text)

The Amazon Resource Name (ARN) for the theme.

updateThemeResponse_creationStatus :: Lens' UpdateThemeResponse (Maybe ResourceStatus)

The creation status of the theme.

updateThemeResponse_versionArn :: Lens' UpdateThemeResponse (Maybe Text)

The Amazon Resource Name (ARN) for the new version of the theme.

updateThemeResponse_status :: Lens' UpdateThemeResponse Int

The HTTP status of the request.