Amazonka.QuickSight.UpdateThemeAlias

Description

Updates an alias of a theme.

Creating a Request

data UpdateThemeAlias

See: newUpdateThemeAlias smart constructor.

Constructors

UpdateThemeAlias'

Fields

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the theme alias that you're updating.

themeId :: Text

The ID for the theme.

aliasName :: Text

The name of the theme alias that you want to update.

themeVersionNumber :: Natural

The version number of the theme that the alias should reference.

newUpdateThemeAlias

Arguments

:: Text
$sel:awsAccountId:UpdateThemeAlias'

-> Text
$sel:themeId:UpdateThemeAlias'

-> Text
$sel:aliasName:UpdateThemeAlias'

-> Natural
$sel:themeVersionNumber:UpdateThemeAlias'

-> UpdateThemeAlias

Create a value of UpdateThemeAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:UpdateThemeAlias', updateThemeAlias_awsAccountId - The ID of the Amazon Web Services account that contains the theme alias that you're updating.

$sel:themeId:UpdateThemeAlias', updateThemeAlias_themeId - The ID for the theme.

$sel:aliasName:UpdateThemeAlias', updateThemeAlias_aliasName - The name of the theme alias that you want to update.

$sel:themeVersionNumber:UpdateThemeAlias', updateThemeAlias_themeVersionNumber - The version number of the theme that the alias should reference.

Request Lenses

updateThemeAlias_awsAccountId :: Lens' UpdateThemeAlias Text

The ID of the Amazon Web Services account that contains the theme alias that you're updating.

updateThemeAlias_themeId :: Lens' UpdateThemeAlias Text

The ID for the theme.

updateThemeAlias_aliasName :: Lens' UpdateThemeAlias Text

The name of the theme alias that you want to update.

updateThemeAlias_themeVersionNumber :: Lens' UpdateThemeAlias Natural

The version number of the theme that the alias should reference.

Destructuring the Response

data UpdateThemeAliasResponse

See: newUpdateThemeAliasResponse smart constructor.

Constructors

UpdateThemeAliasResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

themeAlias :: Maybe ThemeAlias

Information about the theme alias.

status :: Int

The HTTP status of the request.

newUpdateThemeAliasResponse

Arguments

:: Int
$sel:status:UpdateThemeAliasResponse'

-> UpdateThemeAliasResponse

Create a value of UpdateThemeAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateThemeAliasResponse', updateThemeAliasResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:themeAlias:UpdateThemeAliasResponse', updateThemeAliasResponse_themeAlias - Information about the theme alias.

$sel:status:UpdateThemeAliasResponse', updateThemeAliasResponse_status - The HTTP status of the request.

Response Lenses

updateThemeAliasResponse_requestId :: Lens' UpdateThemeAliasResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

updateThemeAliasResponse_themeAlias :: Lens' UpdateThemeAliasResponse (Maybe ThemeAlias)

Information about the theme alias.

updateThemeAliasResponse_status :: Lens' UpdateThemeAliasResponse Int

The HTTP status of the request.