Amazonka.QuickSight.UpdateUser

Description

Updates an Amazon QuickSight user.

Creating a Request

data UpdateUser

See: newUpdateUser smart constructor.

Constructors

UpdateUser'

Fields

unapplyCustomPermissions :: Maybe Bool

A flag that you use to indicate that you want to remove all custom permissions from this user. Using this parameter resets the user to the state it was in before a custom permissions profile was applied. This parameter defaults to NULL and it doesn't accept any other value.

customPermissionsName :: Maybe Text

(Enterprise edition only) The name of the custom permissions profile that you want to assign to this user. Customized permissions allows you to control a user's access by restricting access the following operations:

Create and update data sources

Create and update datasets

Create and update email reports

Subscribe to email reports

A set of custom permissions includes any combination of these restrictions. Currently, you need to create the profile names for custom permission sets by using the Amazon QuickSight console. Then, you use the RegisterUser API operation to assign the named set of permissions to a Amazon QuickSight user.

Amazon QuickSight custom permissions are applied through IAMpolicies. Therefore, they override the permissions typically granted by assigning Amazon QuickSight users to one of the default security cohorts in Amazon QuickSight (admin, author, reader).

This feature is available only to Amazon QuickSight Enterprise edition subscriptions.

customFederationProviderUrl :: Maybe Text

The URL of the custom OpenID Connect (OIDC) provider that provides identity to let a user federate into Amazon QuickSight with an associated Identity and Access Management(IAM) role. This parameter should only be used when ExternalLoginFederationProviderType parameter is set to CUSTOM_OIDC.

externalLoginFederationProviderType :: Maybe Text

The type of supported external login provider that provides identity to let a user federate into Amazon QuickSight with an associated Identity and Access Management(IAM) role. The type of supported external login provider can be one of the following.

COGNITO: Amazon Cognito. The provider URL is cognito-identity.amazonaws.com. When choosing the COGNITO provider type, don’t use the "CustomFederationProviderUrl" parameter which is only needed when the external provider is custom.

CUSTOM_OIDC: Custom OpenID Connect (OIDC) provider. When choosing CUSTOM_OIDC type, use the CustomFederationProviderUrl parameter to provide the custom OIDC provider URL.

NONE: This clears all the previously saved external login information for a user. Use DescribeUser API to check the external login information.

externalLoginId :: Maybe Text

The identity ID for a user in the external login provider.

userName :: Text

The Amazon QuickSight user name that you want to update.

awsAccountId :: Text

The ID for the Amazon Web Services account that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

namespace :: Text

The namespace. Currently, you should set this to default.

email :: Text

The email address of the user that you want to update.

role' :: UserRole

The Amazon QuickSight role of the user. The role can be one of the following default security cohorts:

READER: A user who has read-only access to dashboards.

AUTHOR: A user who can create data sources, datasets, analyses, and dashboards.

ADMIN: A user who is an author, who can also manage Amazon QuickSight settings.

The name of the Amazon QuickSight role is invisible to the user except for the console screens dealing with permissions.

newUpdateUser

Arguments

:: Text
$sel:userName:UpdateUser'

-> Text
$sel:awsAccountId:UpdateUser'

-> Text
$sel:namespace:UpdateUser'

-> Text
$sel:email:UpdateUser'

-> UserRole
$sel:role':UpdateUser'

-> UpdateUser

Create a value of UpdateUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unapplyCustomPermissions:UpdateUser', updateUser_unapplyCustomPermissions - A flag that you use to indicate that you want to remove all custom permissions from this user. Using this parameter resets the user to the state it was in before a custom permissions profile was applied. This parameter defaults to NULL and it doesn't accept any other value.

$sel:customPermissionsName:UpdateUser', updateUser_customPermissionsName - (Enterprise edition only) The name of the custom permissions profile that you want to assign to this user. Customized permissions allows you to control a user's access by restricting access the following operations:

Create and update data sources

Create and update datasets

Create and update email reports

Subscribe to email reports

A set of custom permissions includes any combination of these restrictions. Currently, you need to create the profile names for custom permission sets by using the Amazon QuickSight console. Then, you use the RegisterUser API operation to assign the named set of permissions to a Amazon QuickSight user.

Amazon QuickSight custom permissions are applied through IAMpolicies. Therefore, they override the permissions typically granted by assigning Amazon QuickSight users to one of the default security cohorts in Amazon QuickSight (admin, author, reader).

This feature is available only to Amazon QuickSight Enterprise edition subscriptions.

$sel:customFederationProviderUrl:UpdateUser', updateUser_customFederationProviderUrl - The URL of the custom OpenID Connect (OIDC) provider that provides identity to let a user federate into Amazon QuickSight with an associated Identity and Access Management(IAM) role. This parameter should only be used when ExternalLoginFederationProviderType parameter is set to CUSTOM_OIDC.

$sel:externalLoginFederationProviderType:UpdateUser', updateUser_externalLoginFederationProviderType - The type of supported external login provider that provides identity to let a user federate into Amazon QuickSight with an associated Identity and Access Management(IAM) role. The type of supported external login provider can be one of the following.

COGNITO: Amazon Cognito. The provider URL is cognito-identity.amazonaws.com. When choosing the COGNITO provider type, don’t use the "CustomFederationProviderUrl" parameter which is only needed when the external provider is custom.

CUSTOM_OIDC: Custom OpenID Connect (OIDC) provider. When choosing CUSTOM_OIDC type, use the CustomFederationProviderUrl parameter to provide the custom OIDC provider URL.

NONE: This clears all the previously saved external login information for a user. Use DescribeUser API to check the external login information.

$sel:externalLoginId:UpdateUser', updateUser_externalLoginId - The identity ID for a user in the external login provider.

$sel:userName:UpdateUser', updateUser_userName - The Amazon QuickSight user name that you want to update.

$sel:awsAccountId:UpdateUser', updateUser_awsAccountId - The ID for the Amazon Web Services account that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

$sel:namespace:UpdateUser', updateUser_namespace - The namespace. Currently, you should set this to default.

$sel:email:UpdateUser', updateUser_email - The email address of the user that you want to update.

$sel:role':UpdateUser', updateUser_role - The Amazon QuickSight role of the user. The role can be one of the following default security cohorts:

READER: A user who has read-only access to dashboards.

AUTHOR: A user who can create data sources, datasets, analyses, and dashboards.

ADMIN: A user who is an author, who can also manage Amazon QuickSight settings.

The name of the Amazon QuickSight role is invisible to the user except for the console screens dealing with permissions.

Request Lenses

updateUser_unapplyCustomPermissions :: Lens' UpdateUser (Maybe Bool)

A flag that you use to indicate that you want to remove all custom permissions from this user. Using this parameter resets the user to the state it was in before a custom permissions profile was applied. This parameter defaults to NULL and it doesn't accept any other value.

updateUser_customPermissionsName :: Lens' UpdateUser (Maybe Text)

(Enterprise edition only) The name of the custom permissions profile that you want to assign to this user. Customized permissions allows you to control a user's access by restricting access the following operations:

Create and update data sources

Create and update datasets

Create and update email reports

Subscribe to email reports

A set of custom permissions includes any combination of these restrictions. Currently, you need to create the profile names for custom permission sets by using the Amazon QuickSight console. Then, you use the RegisterUser API operation to assign the named set of permissions to a Amazon QuickSight user.

Amazon QuickSight custom permissions are applied through IAMpolicies. Therefore, they override the permissions typically granted by assigning Amazon QuickSight users to one of the default security cohorts in Amazon QuickSight (admin, author, reader).

This feature is available only to Amazon QuickSight Enterprise edition subscriptions.

updateUser_customFederationProviderUrl :: Lens' UpdateUser (Maybe Text)

The URL of the custom OpenID Connect (OIDC) provider that provides identity to let a user federate into Amazon QuickSight with an associated Identity and Access Management(IAM) role. This parameter should only be used when ExternalLoginFederationProviderType parameter is set to CUSTOM_OIDC.

updateUser_externalLoginFederationProviderType :: Lens' UpdateUser (Maybe Text)

The type of supported external login provider that provides identity to let a user federate into Amazon QuickSight with an associated Identity and Access Management(IAM) role. The type of supported external login provider can be one of the following.

COGNITO: Amazon Cognito. The provider URL is cognito-identity.amazonaws.com. When choosing the COGNITO provider type, don’t use the "CustomFederationProviderUrl" parameter which is only needed when the external provider is custom.

CUSTOM_OIDC: Custom OpenID Connect (OIDC) provider. When choosing CUSTOM_OIDC type, use the CustomFederationProviderUrl parameter to provide the custom OIDC provider URL.

NONE: This clears all the previously saved external login information for a user. Use DescribeUser API to check the external login information.

updateUser_externalLoginId :: Lens' UpdateUser (Maybe Text)

The identity ID for a user in the external login provider.

updateUser_userName :: Lens' UpdateUser Text

The Amazon QuickSight user name that you want to update.

updateUser_awsAccountId :: Lens' UpdateUser Text

The ID for the Amazon Web Services account that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

updateUser_namespace :: Lens' UpdateUser Text

The namespace. Currently, you should set this to default.

updateUser_email :: Lens' UpdateUser Text

The email address of the user that you want to update.

updateUser_role :: Lens' UpdateUser UserRole

The Amazon QuickSight role of the user. The role can be one of the following default security cohorts:

READER: A user who has read-only access to dashboards.

AUTHOR: A user who can create data sources, datasets, analyses, and dashboards.

ADMIN: A user who is an author, who can also manage Amazon QuickSight settings.

The name of the Amazon QuickSight role is invisible to the user except for the console screens dealing with permissions.

Destructuring the Response

data UpdateUserResponse

See: newUpdateUserResponse smart constructor.

Constructors

UpdateUserResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

user :: Maybe User

The Amazon QuickSight user.

status :: Int

The HTTP status of the request.

newUpdateUserResponse

Arguments

:: Int
$sel:status:UpdateUserResponse'

-> UpdateUserResponse

Create a value of UpdateUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateUserResponse', updateUserResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:user:UpdateUserResponse', updateUserResponse_user - The Amazon QuickSight user.

$sel:status:UpdateUserResponse', updateUserResponse_status - The HTTP status of the request.

Response Lenses

updateUserResponse_requestId :: Lens' UpdateUserResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

updateUserResponse_user :: Lens' UpdateUserResponse (Maybe User)

The Amazon QuickSight user.

updateUserResponse_status :: Lens' UpdateUserResponse Int

The HTTP status of the request.