Amazonka.RAM.Lens

Description

Operations

PromoteResourceShareCreatedFromPolicy

promoteResourceShareCreatedFromPolicy_resourceShareArn :: Lens' PromoteResourceShareCreatedFromPolicy Text

The Amazon Resource Name (ARN) of the resource share to promote.

promoteResourceShareCreatedFromPolicyResponse_returnValue :: Lens' PromoteResourceShareCreatedFromPolicyResponse (Maybe Bool)

Indicates whether the request succeeded.

promoteResourceShareCreatedFromPolicyResponse_httpStatus :: Lens' PromoteResourceShareCreatedFromPolicyResponse Int

The response's http status code.

GetResourceShares

getResourceShares_tagFilters :: Lens' GetResourceShares (Maybe [TagFilter])

One or more tag filters.

getResourceShares_nextToken :: Lens' GetResourceShares (Maybe Text)

The token for the next page of results.

getResourceShares_name :: Lens' GetResourceShares (Maybe Text)

The name of the resource share.

getResourceShares_resourceShareStatus :: Lens' GetResourceShares (Maybe ResourceShareStatus)

The status of the resource share.

getResourceShares_permissionArn :: Lens' GetResourceShares (Maybe Text)

The Amazon Resource Name (ARN) of the RAM permission that is associated with the resource share.

getResourceShares_maxResults :: Lens' GetResourceShares (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

getResourceShares_resourceShareArns :: Lens' GetResourceShares (Maybe [Text])

The Amazon Resource Names (ARNs) of the resource shares.

getResourceShares_resourceOwner :: Lens' GetResourceShares ResourceOwner

The type of owner.

getResourceSharesResponse_resourceShares :: Lens' GetResourceSharesResponse (Maybe [ResourceShare])

Information about the resource shares.

getResourceSharesResponse_nextToken :: Lens' GetResourceSharesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

getResourceSharesResponse_httpStatus :: Lens' GetResourceSharesResponse Int

The response's http status code.

GetResourcePolicies

getResourcePolicies_nextToken :: Lens' GetResourcePolicies (Maybe Text)

The token for the next page of results.

getResourcePolicies_principal :: Lens' GetResourcePolicies (Maybe Text)

The principal.

getResourcePolicies_maxResults :: Lens' GetResourcePolicies (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

getResourcePolicies_resourceArns :: Lens' GetResourcePolicies [Text]

The Amazon Resource Names (ARNs) of the resources.

getResourcePoliciesResponse_nextToken :: Lens' GetResourcePoliciesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

getResourcePoliciesResponse_policies :: Lens' GetResourcePoliciesResponse (Maybe [Text])

A key policy document, in JSON format.

getResourcePoliciesResponse_httpStatus :: Lens' GetResourcePoliciesResponse Int

The response's http status code.

ListPendingInvitationResources

listPendingInvitationResources_nextToken :: Lens' ListPendingInvitationResources (Maybe Text)

The token for the next page of results.

listPendingInvitationResources_maxResults :: Lens' ListPendingInvitationResources (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listPendingInvitationResources_resourceShareInvitationArn :: Lens' ListPendingInvitationResources Text

The Amazon Resource Name (ARN) of the invitation.

listPendingInvitationResourcesResponse_resources :: Lens' ListPendingInvitationResourcesResponse (Maybe [Resource])

Information about the resources included the resource share.

listPendingInvitationResourcesResponse_nextToken :: Lens' ListPendingInvitationResourcesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listPendingInvitationResourcesResponse_httpStatus :: Lens' ListPendingInvitationResourcesResponse Int

The response's http status code.

GetPermission

getPermission_permissionVersion :: Lens' GetPermission (Maybe Int)

The identifier for the version of the permission.

getPermission_permissionArn :: Lens' GetPermission Text

The Amazon Resource Name (ARN) of the permission.

getPermissionResponse_permission :: Lens' GetPermissionResponse (Maybe ResourceSharePermissionDetail)

Information about the permission.

getPermissionResponse_httpStatus :: Lens' GetPermissionResponse Int

The response's http status code.

EnableSharingWithAwsOrganization

enableSharingWithAwsOrganizationResponse_returnValue :: Lens' EnableSharingWithAwsOrganizationResponse (Maybe Bool)

Indicates whether the request succeeded.

enableSharingWithAwsOrganizationResponse_httpStatus :: Lens' EnableSharingWithAwsOrganizationResponse Int

The response's http status code.

AssociateResourceSharePermission

associateResourceSharePermission_replace :: Lens' AssociateResourceSharePermission (Maybe Bool)

Indicates whether the permission should replace the permissions that are currently associated with the resource share. Use true to replace the current permissions. Use false to add the permission to the current permission.

associateResourceSharePermission_clientToken :: Lens' AssociateResourceSharePermission (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

associateResourceSharePermission_permissionVersion :: Lens' AssociateResourceSharePermission (Maybe Int)

The version of the RAM permissions to associate with the resource share.

associateResourceSharePermission_resourceShareArn :: Lens' AssociateResourceSharePermission Text

The Amazon Resource Name (ARN) of the resource share.

associateResourceSharePermission_permissionArn :: Lens' AssociateResourceSharePermission Text

The Amazon Resource Name (ARN) of the RAM permission to associate with the resource share.

associateResourceSharePermissionResponse_clientToken :: Lens' AssociateResourceSharePermissionResponse (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

associateResourceSharePermissionResponse_returnValue :: Lens' AssociateResourceSharePermissionResponse (Maybe Bool)

Indicates whether the request succeeded.

associateResourceSharePermissionResponse_httpStatus :: Lens' AssociateResourceSharePermissionResponse Int

The response's http status code.

CreateResourceShare

createResourceShare_clientToken :: Lens' CreateResourceShare (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createResourceShare_allowExternalPrincipals :: Lens' CreateResourceShare (Maybe Bool)

Indicates whether principals outside your organization in Organizations can be associated with a resource share.

createResourceShare_principals :: Lens' CreateResourceShare (Maybe [Text])

The principals to associate with the resource share. The possible values are:

An Amazon Web Services account ID

An Amazon Resource Name (ARN) of an organization in Organizations

An ARN of an organizational unit (OU) in Organizations

An ARN of an IAM role

An ARN of an IAM user

Not all resource types can be shared with IAM roles and IAM users. For more information, see Sharing with IAM roles and IAM users in the Resource Access Manager User Guide.

createResourceShare_resourceArns :: Lens' CreateResourceShare (Maybe [Text])

The ARNs of the resources to associate with the resource share.

createResourceShare_permissionArns :: Lens' CreateResourceShare (Maybe [Text])

The Amazon Resource Names (ARNs) of the permissions to associate with the resource share. If you do not specify an ARN for the permission, RAM automatically attaches the default version of the permission for each resource type. Only one permission can be associated with each resource type in a resource share.

createResourceShare_tags :: Lens' CreateResourceShare (Maybe [Tag])

One or more tags.

createResourceShare_name :: Lens' CreateResourceShare Text

The name of the resource share.

createResourceShareResponse_clientToken :: Lens' CreateResourceShareResponse (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createResourceShareResponse_resourceShare :: Lens' CreateResourceShareResponse (Maybe ResourceShare)

Information about the resource share.

createResourceShareResponse_httpStatus :: Lens' CreateResourceShareResponse Int

The response's http status code.

ListPrincipals

listPrincipals_resourceType :: Lens' ListPrincipals (Maybe Text)

The resource type.

Valid values: acm-pca:CertificateAuthority | appmesh:Mesh | codebuild:Project | codebuild:ReportGroup | ec2:CapacityReservation | ec2:DedicatedHost | ec2:LocalGatewayRouteTable | ec2:PrefixList | ec2:Subnet | ec2:TrafficMirrorTarget | ec2:TransitGateway | imagebuilder:Component | imagebuilder:Image | imagebuilder:ImageRecipe | imagebuilder:ContainerRecipe | glue:Catalog | glue:Database | glue:Table | license-manager:LicenseConfiguration I network-firewall:FirewallPolicy | network-firewall:StatefulRuleGroup | network-firewall:StatelessRuleGroup | outposts:Outpost | resource-groups:Group | rds:Cluster | route53resolver:FirewallRuleGroup |route53resolver:ResolverQueryLogConfig | route53resolver:ResolverRule | s3-outposts:Outpost | ssm-contacts:Contact | ssm-incidents:ResponsePlan

listPrincipals_principals :: Lens' ListPrincipals (Maybe [Text])

The principals.

listPrincipals_nextToken :: Lens' ListPrincipals (Maybe Text)

The token for the next page of results.

listPrincipals_resourceArn :: Lens' ListPrincipals (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

listPrincipals_maxResults :: Lens' ListPrincipals (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listPrincipals_resourceShareArns :: Lens' ListPrincipals (Maybe [Text])

The Amazon Resource Names (ARN) of the resource shares.

listPrincipals_resourceOwner :: Lens' ListPrincipals ResourceOwner

The type of owner.

listPrincipalsResponse_principals :: Lens' ListPrincipalsResponse (Maybe [Principal])

The principals.

listPrincipalsResponse_nextToken :: Lens' ListPrincipalsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listPrincipalsResponse_httpStatus :: Lens' ListPrincipalsResponse Int

The response's http status code.

ListResources

listResources_resourceType :: Lens' ListResources (Maybe Text)

The resource type.

Valid values: acm-pca:CertificateAuthority | appmesh:Mesh | codebuild:Project | codebuild:ReportGroup | ec2:CapacityReservation | ec2:DedicatedHost | ec2:LocalGatewayRouteTable | ec2:PrefixList | ec2:Subnet | ec2:TrafficMirrorTarget | ec2:TransitGateway | imagebuilder:Component | imagebuilder:Image | imagebuilder:ImageRecipe | imagebuilder:ContainerRecipe | glue:Catalog | glue:Database | glue:Table | license-manager:LicenseConfiguration I network-firewall:FirewallPolicy | network-firewall:StatefulRuleGroup | network-firewall:StatelessRuleGroup | outposts:Outpost | resource-groups:Group | rds:Cluster | route53resolver:FirewallRuleGroup |route53resolver:ResolverQueryLogConfig | route53resolver:ResolverRule | s3-outposts:Outpost | ssm-contacts:Contact | ssm-incidents:ResponsePlan

listResources_nextToken :: Lens' ListResources (Maybe Text)

The token for the next page of results.

listResources_resourceArns :: Lens' ListResources (Maybe [Text])

The Amazon Resource Names (ARNs) of the resources.

listResources_principal :: Lens' ListResources (Maybe Text)

The principal.

listResources_maxResults :: Lens' ListResources (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listResources_resourceShareArns :: Lens' ListResources (Maybe [Text])

The Amazon Resource Names (ARN) of the resource shares.

listResources_resourceOwner :: Lens' ListResources ResourceOwner

The type of owner.

listResourcesResponse_resources :: Lens' ListResourcesResponse (Maybe [Resource])

Information about the resources.

listResourcesResponse_nextToken :: Lens' ListResourcesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listResourcesResponse_httpStatus :: Lens' ListResourcesResponse Int

The response's http status code.

AcceptResourceShareInvitation

acceptResourceShareInvitation_clientToken :: Lens' AcceptResourceShareInvitation (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

acceptResourceShareInvitation_resourceShareInvitationArn :: Lens' AcceptResourceShareInvitation Text

The Amazon Resource Name (ARN) of the invitation.

acceptResourceShareInvitationResponse_clientToken :: Lens' AcceptResourceShareInvitationResponse (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

acceptResourceShareInvitationResponse_resourceShareInvitation :: Lens' AcceptResourceShareInvitationResponse (Maybe ResourceShareInvitation)

Information about the invitation.

acceptResourceShareInvitationResponse_httpStatus :: Lens' AcceptResourceShareInvitationResponse Int

The response's http status code.

DeleteResourceShare

deleteResourceShare_clientToken :: Lens' DeleteResourceShare (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

deleteResourceShare_resourceShareArn :: Lens' DeleteResourceShare Text

The Amazon Resource Name (ARN) of the resource share.

deleteResourceShareResponse_clientToken :: Lens' DeleteResourceShareResponse (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

deleteResourceShareResponse_returnValue :: Lens' DeleteResourceShareResponse (Maybe Bool)

Indicates whether the request succeeded.

deleteResourceShareResponse_httpStatus :: Lens' DeleteResourceShareResponse Int

The response's http status code.

UpdateResourceShare

updateResourceShare_clientToken :: Lens' UpdateResourceShare (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

updateResourceShare_allowExternalPrincipals :: Lens' UpdateResourceShare (Maybe Bool)

Indicates whether principals outside your organization in Organizations can be associated with a resource share.

updateResourceShare_name :: Lens' UpdateResourceShare (Maybe Text)

The name of the resource share.

updateResourceShare_resourceShareArn :: Lens' UpdateResourceShare Text

The Amazon Resource Name (ARN) of the resource share.

updateResourceShareResponse_clientToken :: Lens' UpdateResourceShareResponse (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

updateResourceShareResponse_resourceShare :: Lens' UpdateResourceShareResponse (Maybe ResourceShare)

Information about the resource share.

updateResourceShareResponse_httpStatus :: Lens' UpdateResourceShareResponse Int

The response's http status code.

RejectResourceShareInvitation

rejectResourceShareInvitation_clientToken :: Lens' RejectResourceShareInvitation (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

rejectResourceShareInvitation_resourceShareInvitationArn :: Lens' RejectResourceShareInvitation Text

The Amazon Resource Name (ARN) of the invitation.

rejectResourceShareInvitationResponse_clientToken :: Lens' RejectResourceShareInvitationResponse (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

rejectResourceShareInvitationResponse_resourceShareInvitation :: Lens' RejectResourceShareInvitationResponse (Maybe ResourceShareInvitation)

Information about the invitation.

rejectResourceShareInvitationResponse_httpStatus :: Lens' RejectResourceShareInvitationResponse Int

The response's http status code.

ListPermissions

listPermissions_resourceType :: Lens' ListPermissions (Maybe Text)

Specifies the resource type for which to list permissions. For example, to list only permissions that apply to EC2 subnets, specify ec2:Subnet.

listPermissions_nextToken :: Lens' ListPermissions (Maybe Text)

The token for the next page of results.

listPermissions_maxResults :: Lens' ListPermissions (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listPermissionsResponse_nextToken :: Lens' ListPermissionsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listPermissionsResponse_permissions :: Lens' ListPermissionsResponse (Maybe [ResourceSharePermissionSummary])

Information about the permissions.

listPermissionsResponse_httpStatus :: Lens' ListPermissionsResponse Int

The response's http status code.

DisassociateResourceShare

disassociateResourceShare_clientToken :: Lens' DisassociateResourceShare (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

disassociateResourceShare_principals :: Lens' DisassociateResourceShare (Maybe [Text])

The principals.

disassociateResourceShare_resourceArns :: Lens' DisassociateResourceShare (Maybe [Text])

The Amazon Resource Names (ARNs) of the resources.

disassociateResourceShare_resourceShareArn :: Lens' DisassociateResourceShare Text

The Amazon Resource Name (ARN) of the resource share.

disassociateResourceShareResponse_clientToken :: Lens' DisassociateResourceShareResponse (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

disassociateResourceShareResponse_resourceShareAssociations :: Lens' DisassociateResourceShareResponse (Maybe [ResourceShareAssociation])

Information about the associations.

disassociateResourceShareResponse_httpStatus :: Lens' DisassociateResourceShareResponse Int

The response's http status code.

ListResourceSharePermissions

listResourceSharePermissions_nextToken :: Lens' ListResourceSharePermissions (Maybe Text)

The token for the next page of results.

listResourceSharePermissions_maxResults :: Lens' ListResourceSharePermissions (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listResourceSharePermissions_resourceShareArn :: Lens' ListResourceSharePermissions Text

The Amazon Resource Name (ARN) of the resource share.

listResourceSharePermissionsResponse_nextToken :: Lens' ListResourceSharePermissionsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listResourceSharePermissionsResponse_permissions :: Lens' ListResourceSharePermissionsResponse (Maybe [ResourceSharePermissionSummary])

The permissions associated with the resource share.

listResourceSharePermissionsResponse_httpStatus :: Lens' ListResourceSharePermissionsResponse Int

The response's http status code.

TagResource

tagResource_resourceShareArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource share.

tagResource_tags :: Lens' TagResource [Tag]

One or more tags.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListResourceTypes

listResourceTypes_nextToken :: Lens' ListResourceTypes (Maybe Text)

The token for the next page of results.

listResourceTypes_maxResults :: Lens' ListResourceTypes (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listResourceTypesResponse_nextToken :: Lens' ListResourceTypesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listResourceTypesResponse_resourceTypes :: Lens' ListResourceTypesResponse (Maybe [ServiceNameAndResourceType])

The shareable resource types supported by RAM.

listResourceTypesResponse_httpStatus :: Lens' ListResourceTypesResponse Int

The response's http status code.

GetResourceShareAssociations

getResourceShareAssociations_nextToken :: Lens' GetResourceShareAssociations (Maybe Text)

The token for the next page of results.

getResourceShareAssociations_resourceArn :: Lens' GetResourceShareAssociations (Maybe Text)

The Amazon Resource Name (ARN) of the resource. You cannot specify this parameter if the association type is PRINCIPAL.

getResourceShareAssociations_principal :: Lens' GetResourceShareAssociations (Maybe Text)

The principal. You cannot specify this parameter if the association type is RESOURCE.

getResourceShareAssociations_associationStatus :: Lens' GetResourceShareAssociations (Maybe ResourceShareAssociationStatus)

The association status.

getResourceShareAssociations_maxResults :: Lens' GetResourceShareAssociations (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

getResourceShareAssociations_resourceShareArns :: Lens' GetResourceShareAssociations (Maybe [Text])

The Amazon Resource Names (ARN) of the resource shares.

getResourceShareAssociations_associationType :: Lens' GetResourceShareAssociations ResourceShareAssociationType

The association type. Specify PRINCIPAL to list the principals that are associated with the specified resource share. Specify RESOURCE to list the resources that are associated with the specified resource share.

getResourceShareAssociationsResponse_resourceShareAssociations :: Lens' GetResourceShareAssociationsResponse (Maybe [ResourceShareAssociation])

Information about the associations.

getResourceShareAssociationsResponse_nextToken :: Lens' GetResourceShareAssociationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

getResourceShareAssociationsResponse_httpStatus :: Lens' GetResourceShareAssociationsResponse Int

The response's http status code.

GetResourceShareInvitations

getResourceShareInvitations_nextToken :: Lens' GetResourceShareInvitations (Maybe Text)

The token for the next page of results.

getResourceShareInvitations_resourceShareInvitationArns :: Lens' GetResourceShareInvitations (Maybe [Text])

The Amazon Resource Names (ARN) of the invitations.

getResourceShareInvitations_maxResults :: Lens' GetResourceShareInvitations (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

getResourceShareInvitations_resourceShareArns :: Lens' GetResourceShareInvitations (Maybe [Text])

The Amazon Resource Names (ARN) of the resource shares.

getResourceShareInvitationsResponse_resourceShareInvitations :: Lens' GetResourceShareInvitationsResponse (Maybe [ResourceShareInvitation])

Information about the invitations.

getResourceShareInvitationsResponse_nextToken :: Lens' GetResourceShareInvitationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

getResourceShareInvitationsResponse_httpStatus :: Lens' GetResourceShareInvitationsResponse Int

The response's http status code.

UntagResource

untagResource_resourceShareArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource share.

untagResource_tagKeys :: Lens' UntagResource [Text]

The tag keys of the tags to remove.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

AssociateResourceShare

associateResourceShare_clientToken :: Lens' AssociateResourceShare (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

associateResourceShare_principals :: Lens' AssociateResourceShare (Maybe [Text])

The principals to associate with the resource share. The possible values are:

An Amazon Web Services account ID

An Amazon Resource Name (ARN) of an organization in Organizations

An ARN of an organizational unit (OU) in Organizations

An ARN of an IAM role

An ARN of an IAM user

Not all resource types can be shared with IAM roles and IAM users. For more information, see Sharing with IAM roles and IAM users in the Resource Access Manager User Guide.

associateResourceShare_resourceArns :: Lens' AssociateResourceShare (Maybe [Text])

The Amazon Resource Names (ARNs) of the resources.

associateResourceShare_resourceShareArn :: Lens' AssociateResourceShare Text

The Amazon Resource Name (ARN) of the resource share.

associateResourceShareResponse_clientToken :: Lens' AssociateResourceShareResponse (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

associateResourceShareResponse_resourceShareAssociations :: Lens' AssociateResourceShareResponse (Maybe [ResourceShareAssociation])

Information about the associations.

associateResourceShareResponse_httpStatus :: Lens' AssociateResourceShareResponse Int

The response's http status code.

DisassociateResourceSharePermission

disassociateResourceSharePermission_clientToken :: Lens' DisassociateResourceSharePermission (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

disassociateResourceSharePermission_resourceShareArn :: Lens' DisassociateResourceSharePermission Text

The Amazon Resource Name (ARN) of the resource share.

disassociateResourceSharePermission_permissionArn :: Lens' DisassociateResourceSharePermission Text

The Amazon Resource Name (ARN) of the permission to disassociate from the resource share.

disassociateResourceSharePermissionResponse_clientToken :: Lens' DisassociateResourceSharePermissionResponse (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

disassociateResourceSharePermissionResponse_returnValue :: Lens' DisassociateResourceSharePermissionResponse (Maybe Bool)

Indicates whether the request succeeded.

disassociateResourceSharePermissionResponse_httpStatus :: Lens' DisassociateResourceSharePermissionResponse Int

The response's http status code.

Types

Principal

principal_creationTime :: Lens' Principal (Maybe UTCTime)

The time when the principal was associated with the resource share.

principal_resourceShareArn :: Lens' Principal (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

principal_lastUpdatedTime :: Lens' Principal (Maybe UTCTime)

The time when the association was last updated.

principal_external :: Lens' Principal (Maybe Bool)

Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

principal_id :: Lens' Principal (Maybe Text)

The ID of the principal.

Resource

resource_creationTime :: Lens' Resource (Maybe UTCTime)

The time when the resource was associated with the resource share.

resource_status :: Lens' Resource (Maybe ResourceStatus)

The status of the resource.

resource_resourceShareArn :: Lens' Resource (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

resource_lastUpdatedTime :: Lens' Resource (Maybe UTCTime)

The time when the association was last updated.

resource_arn :: Lens' Resource (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

resource_resourceGroupArn :: Lens' Resource (Maybe Text)

The Amazon Resource Name (ARN) of the resource group. This value is returned only if the resource is a resource group.

resource_statusMessage :: Lens' Resource (Maybe Text)

A message about the status of the resource.

resource_type :: Lens' Resource (Maybe Text)

The resource type.

ResourceShare

resourceShare_creationTime :: Lens' ResourceShare (Maybe UTCTime)

The time when the resource share was created.

resourceShare_status :: Lens' ResourceShare (Maybe ResourceShareStatus)

The status of the resource share.

resourceShare_resourceShareArn :: Lens' ResourceShare (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

resourceShare_owningAccountId :: Lens' ResourceShare (Maybe Text)

The ID of the Amazon Web Services account that owns the resource share.

resourceShare_lastUpdatedTime :: Lens' ResourceShare (Maybe UTCTime)

The time when the resource share was last updated.

resourceShare_allowExternalPrincipals :: Lens' ResourceShare (Maybe Bool)

Indicates whether principals outside your organization in Organizations can be associated with a resource share.

resourceShare_name :: Lens' ResourceShare (Maybe Text)

The name of the resource share.

resourceShare_statusMessage :: Lens' ResourceShare (Maybe Text)

A message about the status of the resource share.

resourceShare_featureSet :: Lens' ResourceShare (Maybe ResourceShareFeatureSet)

Indicates how the resource share was created. Possible values include:

CREATED_FROM_POLICY - Indicates that the resource share was created from an Amazon Web Services Identity and Access Management (Amazon Web Services IAM) policy attached to a resource. These resource shares are visible only to the Amazon Web Services account that created it. They cannot be modified in RAM.

PROMOTING_TO_STANDARD - The resource share is in the process of being promoted. For more information, see PromoteResourceShareCreatedFromPolicy.

STANDARD - Indicates that the resource share was created in RAM using the console or APIs. These resource shares are visible to all principals. They can be modified in RAM.

resourceShare_tags :: Lens' ResourceShare (Maybe [Tag])

The tags for the resource share.

ResourceShareAssociation

resourceShareAssociation_creationTime :: Lens' ResourceShareAssociation (Maybe UTCTime)

The time when the association was created.

resourceShareAssociation_status :: Lens' ResourceShareAssociation (Maybe ResourceShareAssociationStatus)

The status of the association.

resourceShareAssociation_resourceShareArn :: Lens' ResourceShareAssociation (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

resourceShareAssociation_lastUpdatedTime :: Lens' ResourceShareAssociation (Maybe UTCTime)

The time when the association was last updated.

resourceShareAssociation_external :: Lens' ResourceShareAssociation (Maybe Bool)

Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

resourceShareAssociation_resourceShareName :: Lens' ResourceShareAssociation (Maybe Text)

The name of the resource share.

resourceShareAssociation_associatedEntity :: Lens' ResourceShareAssociation (Maybe Text)

The associated entity. For resource associations, this is the Amazon Resource Name (ARN) of the resource. For principal associations, this is one of the following:

An Amazon Web Services account ID

An ARN of an organization in Organizations

An ARN of an organizational unit (OU) in Organizations

An ARN of an IAM role

An ARN of an IAM user

resourceShareAssociation_associationType :: Lens' ResourceShareAssociation (Maybe ResourceShareAssociationType)

The association type.

resourceShareAssociation_statusMessage :: Lens' ResourceShareAssociation (Maybe Text)

A message about the status of the association.

ResourceShareInvitation

resourceShareInvitation_status :: Lens' ResourceShareInvitation (Maybe ResourceShareInvitationStatus)

The status of the invitation.

resourceShareInvitation_senderAccountId :: Lens' ResourceShareInvitation (Maybe Text)

The ID of the Amazon Web Services account that sent the invitation.

resourceShareInvitation_resourceShareArn :: Lens' ResourceShareInvitation (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

resourceShareInvitation_receiverAccountId :: Lens' ResourceShareInvitation (Maybe Text)

The ID of the Amazon Web Services account that received the invitation.

resourceShareInvitation_resourceShareAssociations :: Lens' ResourceShareInvitation (Maybe [ResourceShareAssociation])

To view the resources associated with a pending resource share invitation, use ListPendingInvitationResources.

resourceShareInvitation_resourceShareName :: Lens' ResourceShareInvitation (Maybe Text)

The name of the resource share.

resourceShareInvitation_receiverArn :: Lens' ResourceShareInvitation (Maybe Text)

The Amazon Resource Name (ARN) of the IAM user or IAM role that received the invitation.

resourceShareInvitation_invitationTimestamp :: Lens' ResourceShareInvitation (Maybe UTCTime)

The date and time when the invitation was sent.

resourceShareInvitation_resourceShareInvitationArn :: Lens' ResourceShareInvitation (Maybe Text)

The Amazon Resource Name (ARN) of the invitation.

ResourceSharePermissionDetail

resourceSharePermissionDetail_creationTime :: Lens' ResourceSharePermissionDetail (Maybe UTCTime)

The date and time when the permission was created.

resourceSharePermissionDetail_resourceType :: Lens' ResourceSharePermissionDetail (Maybe Text)

The resource type to which the permission applies.

resourceSharePermissionDetail_lastUpdatedTime :: Lens' ResourceSharePermissionDetail (Maybe UTCTime)

The date and time when the permission was last updated.

resourceSharePermissionDetail_arn :: Lens' ResourceSharePermissionDetail (Maybe Text)

The Amazon Resource Name (ARN) of the permission.

resourceSharePermissionDetail_defaultVersion :: Lens' ResourceSharePermissionDetail (Maybe Bool)

Specifies whether the version of the permission is set to the default version for this permission.

resourceSharePermissionDetail_name :: Lens' ResourceSharePermissionDetail (Maybe Text)

The name of the permission.

resourceSharePermissionDetail_version :: Lens' ResourceSharePermissionDetail (Maybe Text)

The identifier for the version of the permission.

resourceSharePermissionDetail_isResourceTypeDefault :: Lens' ResourceSharePermissionDetail (Maybe Bool)

Specifies whether the version of the permission is set to the default version for this resource type.

resourceSharePermissionDetail_permission :: Lens' ResourceSharePermissionDetail (Maybe Text)

The permission's effect and actions in JSON format. The effect indicates whether the actions are allowed or denied. The actions list the API actions to which the principal is granted or denied access.

ResourceSharePermissionSummary

resourceSharePermissionSummary_creationTime :: Lens' ResourceSharePermissionSummary (Maybe UTCTime)

The date and time when the permission was created.

resourceSharePermissionSummary_status :: Lens' ResourceSharePermissionSummary (Maybe Text)

The current status of the permission.

resourceSharePermissionSummary_resourceType :: Lens' ResourceSharePermissionSummary (Maybe Text)

The type of resource to which the permission applies.

resourceSharePermissionSummary_lastUpdatedTime :: Lens' ResourceSharePermissionSummary (Maybe UTCTime)

The date and time when the permission was last updated.

resourceSharePermissionSummary_arn :: Lens' ResourceSharePermissionSummary (Maybe Text)

The Amazon Resource Name (ARN) of the permission.

resourceSharePermissionSummary_defaultVersion :: Lens' ResourceSharePermissionSummary (Maybe Bool)

Specifies whether the version of the permission is set to the default version for this permission.

resourceSharePermissionSummary_name :: Lens' ResourceSharePermissionSummary (Maybe Text)

The name of the permission.

resourceSharePermissionSummary_version :: Lens' ResourceSharePermissionSummary (Maybe Text)

The identifier for the version of the permission.

resourceSharePermissionSummary_isResourceTypeDefault :: Lens' ResourceSharePermissionSummary (Maybe Bool)

Specifies whether the version of the permission is set to the default version for this resource type.

ServiceNameAndResourceType

serviceNameAndResourceType_resourceType :: Lens' ServiceNameAndResourceType (Maybe Text)

The shareable resource types.

serviceNameAndResourceType_serviceName :: Lens' ServiceNameAndResourceType (Maybe Text)

The name of the Amazon Web Services services to which the resources belong.

Tag

tag_value :: Lens' Tag (Maybe Text)

The value of the tag.

tag_key :: Lens' Tag (Maybe Text)

The key of the tag.

TagFilter

tagFilter_tagValues :: Lens' TagFilter (Maybe [Text])

The tag values.

tagFilter_tagKey :: Lens' TagFilter (Maybe Text)

The tag key.