Amazonka.RAM.ListPendingInvitationResources

Description

Lists the resources in a resource share that is shared with you but that the invitation is still pending for.

Creating a Request

data ListPendingInvitationResources

See: newListPendingInvitationResources smart constructor.

Constructors

ListPendingInvitationResources'

Fields

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

resourceShareInvitationArn :: Text

The Amazon Resource Name (ARN) of the invitation.

newListPendingInvitationResources

Arguments

:: Text
$sel:resourceShareInvitationArn:ListPendingInvitationResources'

-> ListPendingInvitationResources

Create a value of ListPendingInvitationResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPendingInvitationResources', listPendingInvitationResources_nextToken - The token for the next page of results.

$sel:maxResults:ListPendingInvitationResources', listPendingInvitationResources_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:resourceShareInvitationArn:ListPendingInvitationResources', listPendingInvitationResources_resourceShareInvitationArn - The Amazon Resource Name (ARN) of the invitation.

Request Lenses

listPendingInvitationResources_nextToken :: Lens' ListPendingInvitationResources (Maybe Text)

The token for the next page of results.

listPendingInvitationResources_maxResults :: Lens' ListPendingInvitationResources (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listPendingInvitationResources_resourceShareInvitationArn :: Lens' ListPendingInvitationResources Text

The Amazon Resource Name (ARN) of the invitation.

Destructuring the Response

data ListPendingInvitationResourcesResponse

See: newListPendingInvitationResourcesResponse smart constructor.

Constructors

ListPendingInvitationResourcesResponse'

Fields

resources :: Maybe [Resource]

Information about the resources included the resource share.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newListPendingInvitationResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListPendingInvitationResourcesResponse'

-> ListPendingInvitationResourcesResponse

Create a value of ListPendingInvitationResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resources:ListPendingInvitationResourcesResponse', listPendingInvitationResourcesResponse_resources - Information about the resources included the resource share.

$sel:nextToken:ListPendingInvitationResources', listPendingInvitationResourcesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListPendingInvitationResourcesResponse', listPendingInvitationResourcesResponse_httpStatus - The response's http status code.

Response Lenses

listPendingInvitationResourcesResponse_resources :: Lens' ListPendingInvitationResourcesResponse (Maybe [Resource])

Information about the resources included the resource share.

listPendingInvitationResourcesResponse_nextToken :: Lens' ListPendingInvitationResourcesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listPendingInvitationResourcesResponse_httpStatus :: Lens' ListPendingInvitationResourcesResponse Int

The response's http status code.