Amazonka.RAM.ListResources

Description

Lists the resources that you added to a resource shares or the resources that are shared with you.

This operation returns paginated results.

Creating a Request

data ListResources

See: newListResources smart constructor.

Constructors

ListResources'

Fields

resourceType :: Maybe Text

The resource type.

Valid values: acm-pca:CertificateAuthority | appmesh:Mesh | codebuild:Project | codebuild:ReportGroup | ec2:CapacityReservation | ec2:DedicatedHost | ec2:LocalGatewayRouteTable | ec2:PrefixList | ec2:Subnet | ec2:TrafficMirrorTarget | ec2:TransitGateway | imagebuilder:Component | imagebuilder:Image | imagebuilder:ImageRecipe | imagebuilder:ContainerRecipe | glue:Catalog | glue:Database | glue:Table | license-manager:LicenseConfiguration I network-firewall:FirewallPolicy | network-firewall:StatefulRuleGroup | network-firewall:StatelessRuleGroup | outposts:Outpost | resource-groups:Group | rds:Cluster | route53resolver:FirewallRuleGroup |route53resolver:ResolverQueryLogConfig | route53resolver:ResolverRule | s3-outposts:Outpost | ssm-contacts:Contact | ssm-incidents:ResponsePlan

nextToken :: Maybe Text

The token for the next page of results.

resourceArns :: Maybe [Text]

The Amazon Resource Names (ARNs) of the resources.

principal :: Maybe Text

The principal.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

resourceShareArns :: Maybe [Text]

The Amazon Resource Names (ARN) of the resource shares.

resourceOwner :: ResourceOwner

The type of owner.

newListResources

Arguments

:: ResourceOwner
$sel:resourceOwner:ListResources'

-> ListResources

Create a value of ListResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ListResources', listResources_resourceType - The resource type.

Valid values: acm-pca:CertificateAuthority | appmesh:Mesh | codebuild:Project | codebuild:ReportGroup | ec2:CapacityReservation | ec2:DedicatedHost | ec2:LocalGatewayRouteTable | ec2:PrefixList | ec2:Subnet | ec2:TrafficMirrorTarget | ec2:TransitGateway | imagebuilder:Component | imagebuilder:Image | imagebuilder:ImageRecipe | imagebuilder:ContainerRecipe | glue:Catalog | glue:Database | glue:Table | license-manager:LicenseConfiguration I network-firewall:FirewallPolicy | network-firewall:StatefulRuleGroup | network-firewall:StatelessRuleGroup | outposts:Outpost | resource-groups:Group | rds:Cluster | route53resolver:FirewallRuleGroup |route53resolver:ResolverQueryLogConfig | route53resolver:ResolverRule | s3-outposts:Outpost | ssm-contacts:Contact | ssm-incidents:ResponsePlan

$sel:nextToken:ListResources', listResources_nextToken - The token for the next page of results.

$sel:resourceArns:ListResources', listResources_resourceArns - The Amazon Resource Names (ARNs) of the resources.

$sel:principal:ListResources', listResources_principal - The principal.

$sel:maxResults:ListResources', listResources_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:resourceShareArns:ListResources', listResources_resourceShareArns - The Amazon Resource Names (ARN) of the resource shares.

$sel:resourceOwner:ListResources', listResources_resourceOwner - The type of owner.

Request Lenses

listResources_resourceType :: Lens' ListResources (Maybe Text)

The resource type.

Valid values: acm-pca:CertificateAuthority | appmesh:Mesh | codebuild:Project | codebuild:ReportGroup | ec2:CapacityReservation | ec2:DedicatedHost | ec2:LocalGatewayRouteTable | ec2:PrefixList | ec2:Subnet | ec2:TrafficMirrorTarget | ec2:TransitGateway | imagebuilder:Component | imagebuilder:Image | imagebuilder:ImageRecipe | imagebuilder:ContainerRecipe | glue:Catalog | glue:Database | glue:Table | license-manager:LicenseConfiguration I network-firewall:FirewallPolicy | network-firewall:StatefulRuleGroup | network-firewall:StatelessRuleGroup | outposts:Outpost | resource-groups:Group | rds:Cluster | route53resolver:FirewallRuleGroup |route53resolver:ResolverQueryLogConfig | route53resolver:ResolverRule | s3-outposts:Outpost | ssm-contacts:Contact | ssm-incidents:ResponsePlan

listResources_nextToken :: Lens' ListResources (Maybe Text)

The token for the next page of results.

listResources_resourceArns :: Lens' ListResources (Maybe [Text])

The Amazon Resource Names (ARNs) of the resources.

listResources_principal :: Lens' ListResources (Maybe Text)

The principal.

listResources_maxResults :: Lens' ListResources (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listResources_resourceShareArns :: Lens' ListResources (Maybe [Text])

The Amazon Resource Names (ARN) of the resource shares.

listResources_resourceOwner :: Lens' ListResources ResourceOwner

The type of owner.

Destructuring the Response

data ListResourcesResponse

See: newListResourcesResponse smart constructor.

Constructors

ListResourcesResponse'

Fields

resources :: Maybe [Resource]

Information about the resources.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newListResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListResourcesResponse'

-> ListResourcesResponse

Create a value of ListResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resources:ListResourcesResponse', listResourcesResponse_resources - Information about the resources.

$sel:nextToken:ListResources', listResourcesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListResourcesResponse', listResourcesResponse_httpStatus - The response's http status code.

Response Lenses

listResourcesResponse_resources :: Lens' ListResourcesResponse (Maybe [Resource])

Information about the resources.

listResourcesResponse_nextToken :: Lens' ListResourcesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listResourcesResponse_httpStatus :: Lens' ListResourcesResponse Int

The response's http status code.