Amazonka.RAM.Types.ResourceShare

Description

Documentation

data ResourceShare

Describes a resource share.

See: newResourceShare smart constructor.

Constructors

ResourceShare'

Fields

creationTime :: Maybe POSIX

The time when the resource share was created.

status :: Maybe ResourceShareStatus

The status of the resource share.

resourceShareArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource share.

owningAccountId :: Maybe Text

The ID of the Amazon Web Services account that owns the resource share.

lastUpdatedTime :: Maybe POSIX

The time when the resource share was last updated.

allowExternalPrincipals :: Maybe Bool

Indicates whether principals outside your organization in Organizations can be associated with a resource share.

name :: Maybe Text

The name of the resource share.

statusMessage :: Maybe Text

A message about the status of the resource share.

featureSet :: Maybe ResourceShareFeatureSet

Indicates how the resource share was created. Possible values include:

CREATED_FROM_POLICY - Indicates that the resource share was created from an Amazon Web Services Identity and Access Management (Amazon Web Services IAM) policy attached to a resource. These resource shares are visible only to the Amazon Web Services account that created it. They cannot be modified in RAM.

PROMOTING_TO_STANDARD - The resource share is in the process of being promoted. For more information, see PromoteResourceShareCreatedFromPolicy.

STANDARD - Indicates that the resource share was created in RAM using the console or APIs. These resource shares are visible to all principals. They can be modified in RAM.

tags :: Maybe [Tag]

The tags for the resource share.

newResourceShare :: ResourceShare

Create a value of ResourceShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ResourceShare', resourceShare_creationTime - The time when the resource share was created.

$sel:status:ResourceShare', resourceShare_status - The status of the resource share.

$sel:resourceShareArn:ResourceShare', resourceShare_resourceShareArn - The Amazon Resource Name (ARN) of the resource share.

$sel:owningAccountId:ResourceShare', resourceShare_owningAccountId - The ID of the Amazon Web Services account that owns the resource share.

$sel:lastUpdatedTime:ResourceShare', resourceShare_lastUpdatedTime - The time when the resource share was last updated.

$sel:allowExternalPrincipals:ResourceShare', resourceShare_allowExternalPrincipals - Indicates whether principals outside your organization in Organizations can be associated with a resource share.

$sel:name:ResourceShare', resourceShare_name - The name of the resource share.

$sel:statusMessage:ResourceShare', resourceShare_statusMessage - A message about the status of the resource share.

$sel:featureSet:ResourceShare', resourceShare_featureSet - Indicates how the resource share was created. Possible values include:

CREATED_FROM_POLICY - Indicates that the resource share was created from an Amazon Web Services Identity and Access Management (Amazon Web Services IAM) policy attached to a resource. These resource shares are visible only to the Amazon Web Services account that created it. They cannot be modified in RAM.

PROMOTING_TO_STANDARD - The resource share is in the process of being promoted. For more information, see PromoteResourceShareCreatedFromPolicy.

STANDARD - Indicates that the resource share was created in RAM using the console or APIs. These resource shares are visible to all principals. They can be modified in RAM.

$sel:tags:ResourceShare', resourceShare_tags - The tags for the resource share.

resourceShare_creationTime :: Lens' ResourceShare (Maybe UTCTime)

The time when the resource share was created.

resourceShare_status :: Lens' ResourceShare (Maybe ResourceShareStatus)

The status of the resource share.

resourceShare_resourceShareArn :: Lens' ResourceShare (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

resourceShare_owningAccountId :: Lens' ResourceShare (Maybe Text)

The ID of the Amazon Web Services account that owns the resource share.

resourceShare_lastUpdatedTime :: Lens' ResourceShare (Maybe UTCTime)

The time when the resource share was last updated.

resourceShare_allowExternalPrincipals :: Lens' ResourceShare (Maybe Bool)

Indicates whether principals outside your organization in Organizations can be associated with a resource share.

resourceShare_name :: Lens' ResourceShare (Maybe Text)

The name of the resource share.

resourceShare_statusMessage :: Lens' ResourceShare (Maybe Text)

A message about the status of the resource share.

resourceShare_featureSet :: Lens' ResourceShare (Maybe ResourceShareFeatureSet)

Indicates how the resource share was created. Possible values include:

CREATED_FROM_POLICY - Indicates that the resource share was created from an Amazon Web Services Identity and Access Management (Amazon Web Services IAM) policy attached to a resource. These resource shares are visible only to the Amazon Web Services account that created it. They cannot be modified in RAM.

PROMOTING_TO_STANDARD - The resource share is in the process of being promoted. For more information, see PromoteResourceShareCreatedFromPolicy.

STANDARD - Indicates that the resource share was created in RAM using the console or APIs. These resource shares are visible to all principals. They can be modified in RAM.

resourceShare_tags :: Lens' ResourceShare (Maybe [Tag])

The tags for the resource share.