Amazonka.RAM.Types.ResourceShareAssociation

Description

Documentation

data ResourceShareAssociation

Describes an association with a resource share.

See: newResourceShareAssociation smart constructor.

Constructors

ResourceShareAssociation'

Fields

creationTime :: Maybe POSIX

The time when the association was created.

status :: Maybe ResourceShareAssociationStatus

The status of the association.

resourceShareArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource share.

lastUpdatedTime :: Maybe POSIX

The time when the association was last updated.

external :: Maybe Bool

Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

resourceShareName :: Maybe Text

The name of the resource share.

associatedEntity :: Maybe Text

The associated entity. For resource associations, this is the Amazon Resource Name (ARN) of the resource. For principal associations, this is one of the following:

An Amazon Web Services account ID

An ARN of an organization in Organizations

An ARN of an organizational unit (OU) in Organizations

An ARN of an IAM role

An ARN of an IAM user

associationType :: Maybe ResourceShareAssociationType

The association type.

statusMessage :: Maybe Text

A message about the status of the association.

newResourceShareAssociation :: ResourceShareAssociation

Create a value of ResourceShareAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ResourceShareAssociation', resourceShareAssociation_creationTime - The time when the association was created.

$sel:status:ResourceShareAssociation', resourceShareAssociation_status - The status of the association.

$sel:resourceShareArn:ResourceShareAssociation', resourceShareAssociation_resourceShareArn - The Amazon Resource Name (ARN) of the resource share.

$sel:lastUpdatedTime:ResourceShareAssociation', resourceShareAssociation_lastUpdatedTime - The time when the association was last updated.

$sel:external:ResourceShareAssociation', resourceShareAssociation_external - Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

$sel:resourceShareName:ResourceShareAssociation', resourceShareAssociation_resourceShareName - The name of the resource share.

$sel:associatedEntity:ResourceShareAssociation', resourceShareAssociation_associatedEntity - The associated entity. For resource associations, this is the Amazon Resource Name (ARN) of the resource. For principal associations, this is one of the following:

An Amazon Web Services account ID

An ARN of an organization in Organizations

An ARN of an organizational unit (OU) in Organizations

An ARN of an IAM role

An ARN of an IAM user

$sel:associationType:ResourceShareAssociation', resourceShareAssociation_associationType - The association type.

$sel:statusMessage:ResourceShareAssociation', resourceShareAssociation_statusMessage - A message about the status of the association.

resourceShareAssociation_creationTime :: Lens' ResourceShareAssociation (Maybe UTCTime)

The time when the association was created.

resourceShareAssociation_status :: Lens' ResourceShareAssociation (Maybe ResourceShareAssociationStatus)

The status of the association.

resourceShareAssociation_resourceShareArn :: Lens' ResourceShareAssociation (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

resourceShareAssociation_lastUpdatedTime :: Lens' ResourceShareAssociation (Maybe UTCTime)

The time when the association was last updated.

resourceShareAssociation_external :: Lens' ResourceShareAssociation (Maybe Bool)

Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

resourceShareAssociation_resourceShareName :: Lens' ResourceShareAssociation (Maybe Text)

The name of the resource share.

resourceShareAssociation_associatedEntity :: Lens' ResourceShareAssociation (Maybe Text)

The associated entity. For resource associations, this is the Amazon Resource Name (ARN) of the resource. For principal associations, this is one of the following:

An Amazon Web Services account ID

An ARN of an organization in Organizations

An ARN of an organizational unit (OU) in Organizations

An ARN of an IAM role

An ARN of an IAM user

resourceShareAssociation_associationType :: Lens' ResourceShareAssociation (Maybe ResourceShareAssociationType)

The association type.

resourceShareAssociation_statusMessage :: Lens' ResourceShareAssociation (Maybe Text)

A message about the status of the association.