Amazonka.RAM.Types.ResourceShareInvitation

Description

Documentation

data ResourceShareInvitation

Describes an invitation to join a resource share.

See: newResourceShareInvitation smart constructor.

Constructors

ResourceShareInvitation'

Fields

status :: Maybe ResourceShareInvitationStatus

The status of the invitation.

senderAccountId :: Maybe Text

The ID of the Amazon Web Services account that sent the invitation.

resourceShareArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource share.

receiverAccountId :: Maybe Text

The ID of the Amazon Web Services account that received the invitation.

resourceShareAssociations :: Maybe [ResourceShareAssociation]

To view the resources associated with a pending resource share invitation, use ListPendingInvitationResources.

resourceShareName :: Maybe Text

The name of the resource share.

receiverArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM user or IAM role that received the invitation.

invitationTimestamp :: Maybe POSIX

The date and time when the invitation was sent.

resourceShareInvitationArn :: Maybe Text

The Amazon Resource Name (ARN) of the invitation.

newResourceShareInvitation :: ResourceShareInvitation

Create a value of ResourceShareInvitation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ResourceShareInvitation', resourceShareInvitation_status - The status of the invitation.

$sel:senderAccountId:ResourceShareInvitation', resourceShareInvitation_senderAccountId - The ID of the Amazon Web Services account that sent the invitation.

$sel:resourceShareArn:ResourceShareInvitation', resourceShareInvitation_resourceShareArn - The Amazon Resource Name (ARN) of the resource share.

$sel:receiverAccountId:ResourceShareInvitation', resourceShareInvitation_receiverAccountId - The ID of the Amazon Web Services account that received the invitation.

$sel:resourceShareAssociations:ResourceShareInvitation', resourceShareInvitation_resourceShareAssociations - To view the resources associated with a pending resource share invitation, use ListPendingInvitationResources.

$sel:resourceShareName:ResourceShareInvitation', resourceShareInvitation_resourceShareName - The name of the resource share.

$sel:receiverArn:ResourceShareInvitation', resourceShareInvitation_receiverArn - The Amazon Resource Name (ARN) of the IAM user or IAM role that received the invitation.

$sel:invitationTimestamp:ResourceShareInvitation', resourceShareInvitation_invitationTimestamp - The date and time when the invitation was sent.

$sel:resourceShareInvitationArn:ResourceShareInvitation', resourceShareInvitation_resourceShareInvitationArn - The Amazon Resource Name (ARN) of the invitation.

resourceShareInvitation_status :: Lens' ResourceShareInvitation (Maybe ResourceShareInvitationStatus)

The status of the invitation.

resourceShareInvitation_senderAccountId :: Lens' ResourceShareInvitation (Maybe Text)

The ID of the Amazon Web Services account that sent the invitation.

resourceShareInvitation_resourceShareArn :: Lens' ResourceShareInvitation (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

resourceShareInvitation_receiverAccountId :: Lens' ResourceShareInvitation (Maybe Text)

The ID of the Amazon Web Services account that received the invitation.

resourceShareInvitation_resourceShareAssociations :: Lens' ResourceShareInvitation (Maybe [ResourceShareAssociation])

To view the resources associated with a pending resource share invitation, use ListPendingInvitationResources.

resourceShareInvitation_resourceShareName :: Lens' ResourceShareInvitation (Maybe Text)

The name of the resource share.

resourceShareInvitation_receiverArn :: Lens' ResourceShareInvitation (Maybe Text)

The Amazon Resource Name (ARN) of the IAM user or IAM role that received the invitation.

resourceShareInvitation_invitationTimestamp :: Lens' ResourceShareInvitation (Maybe UTCTime)

The date and time when the invitation was sent.

resourceShareInvitation_resourceShareInvitationArn :: Lens' ResourceShareInvitation (Maybe Text)

The Amazon Resource Name (ARN) of the invitation.