Amazonka.RAM.Types.ResourceSharePermissionSummary

Description

Documentation

data ResourceSharePermissionSummary

Information about a permission that is associated with a resource share.

See: newResourceSharePermissionSummary smart constructor.

Constructors

ResourceSharePermissionSummary'

Fields

creationTime :: Maybe POSIX

The date and time when the permission was created.

status :: Maybe Text

The current status of the permission.

resourceType :: Maybe Text

The type of resource to which the permission applies.

lastUpdatedTime :: Maybe POSIX

The date and time when the permission was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the permission.

defaultVersion :: Maybe Bool

Specifies whether the version of the permission is set to the default version for this permission.

name :: Maybe Text

The name of the permission.

version :: Maybe Text

The identifier for the version of the permission.

isResourceTypeDefault :: Maybe Bool

Specifies whether the version of the permission is set to the default version for this resource type.

newResourceSharePermissionSummary :: ResourceSharePermissionSummary

Create a value of ResourceSharePermissionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ResourceSharePermissionSummary', resourceSharePermissionSummary_creationTime - The date and time when the permission was created.

$sel:status:ResourceSharePermissionSummary', resourceSharePermissionSummary_status - The current status of the permission.

$sel:resourceType:ResourceSharePermissionSummary', resourceSharePermissionSummary_resourceType - The type of resource to which the permission applies.

$sel:lastUpdatedTime:ResourceSharePermissionSummary', resourceSharePermissionSummary_lastUpdatedTime - The date and time when the permission was last updated.

$sel:arn:ResourceSharePermissionSummary', resourceSharePermissionSummary_arn - The Amazon Resource Name (ARN) of the permission.

$sel:defaultVersion:ResourceSharePermissionSummary', resourceSharePermissionSummary_defaultVersion - Specifies whether the version of the permission is set to the default version for this permission.

$sel:name:ResourceSharePermissionSummary', resourceSharePermissionSummary_name - The name of the permission.

$sel:version:ResourceSharePermissionSummary', resourceSharePermissionSummary_version - The identifier for the version of the permission.

$sel:isResourceTypeDefault:ResourceSharePermissionSummary', resourceSharePermissionSummary_isResourceTypeDefault - Specifies whether the version of the permission is set to the default version for this resource type.

resourceSharePermissionSummary_creationTime :: Lens' ResourceSharePermissionSummary (Maybe UTCTime)

The date and time when the permission was created.

resourceSharePermissionSummary_status :: Lens' ResourceSharePermissionSummary (Maybe Text)

The current status of the permission.

resourceSharePermissionSummary_resourceType :: Lens' ResourceSharePermissionSummary (Maybe Text)

The type of resource to which the permission applies.

resourceSharePermissionSummary_lastUpdatedTime :: Lens' ResourceSharePermissionSummary (Maybe UTCTime)

The date and time when the permission was last updated.

resourceSharePermissionSummary_arn :: Lens' ResourceSharePermissionSummary (Maybe Text)

The Amazon Resource Name (ARN) of the permission.

resourceSharePermissionSummary_defaultVersion :: Lens' ResourceSharePermissionSummary (Maybe Bool)

Specifies whether the version of the permission is set to the default version for this permission.

resourceSharePermissionSummary_name :: Lens' ResourceSharePermissionSummary (Maybe Text)

The name of the permission.

resourceSharePermissionSummary_version :: Lens' ResourceSharePermissionSummary (Maybe Text)

The identifier for the version of the permission.

resourceSharePermissionSummary_isResourceTypeDefault :: Lens' ResourceSharePermissionSummary (Maybe Bool)

Specifies whether the version of the permission is set to the default version for this resource type.