Amazonka.RDS.CreateDBClusterEndpoint

Description

Creates a new custom endpoint and associates it with an Amazon Aurora DB cluster.

This action only applies to Aurora DB clusters.

Creating a Request

data CreateDBClusterEndpoint

See: newCreateDBClusterEndpoint smart constructor.

Constructors

CreateDBClusterEndpoint'

Fields

staticMembers :: Maybe [Text]

List of DB instance identifiers that are part of the custom endpoint group.

excludedMembers :: Maybe [Text]

List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

tags :: Maybe [Tag]

The tags to be assigned to the Amazon RDS resource.

dbClusterIdentifier :: Text

The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

dbClusterEndpointIdentifier :: Text

The identifier to use for the new endpoint. This parameter is stored as a lowercase string.

endpointType :: Text

The type of the endpoint. One of: READER, WRITER, ANY.

newCreateDBClusterEndpoint

Arguments

:: Text
$sel:dbClusterIdentifier:CreateDBClusterEndpoint'

-> Text
$sel:dbClusterEndpointIdentifier:CreateDBClusterEndpoint'

-> Text
$sel:endpointType:CreateDBClusterEndpoint'

-> CreateDBClusterEndpoint

Create a value of CreateDBClusterEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:staticMembers:CreateDBClusterEndpoint', createDBClusterEndpoint_staticMembers - List of DB instance identifiers that are part of the custom endpoint group.

$sel:excludedMembers:CreateDBClusterEndpoint', createDBClusterEndpoint_excludedMembers - List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

$sel:tags:CreateDBClusterEndpoint', createDBClusterEndpoint_tags - The tags to be assigned to the Amazon RDS resource.

$sel:dbClusterIdentifier:CreateDBClusterEndpoint', createDBClusterEndpoint_dbClusterIdentifier - The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

$sel:dbClusterEndpointIdentifier:CreateDBClusterEndpoint', createDBClusterEndpoint_dbClusterEndpointIdentifier - The identifier to use for the new endpoint. This parameter is stored as a lowercase string.

$sel:endpointType:CreateDBClusterEndpoint', createDBClusterEndpoint_endpointType - The type of the endpoint. One of: READER, WRITER, ANY.

Request Lenses

createDBClusterEndpoint_staticMembers :: Lens' CreateDBClusterEndpoint (Maybe [Text])

List of DB instance identifiers that are part of the custom endpoint group.

createDBClusterEndpoint_excludedMembers :: Lens' CreateDBClusterEndpoint (Maybe [Text])

List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

createDBClusterEndpoint_tags :: Lens' CreateDBClusterEndpoint (Maybe [Tag])

The tags to be assigned to the Amazon RDS resource.

createDBClusterEndpoint_dbClusterIdentifier :: Lens' CreateDBClusterEndpoint Text

The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

createDBClusterEndpoint_dbClusterEndpointIdentifier :: Lens' CreateDBClusterEndpoint Text

The identifier to use for the new endpoint. This parameter is stored as a lowercase string.

createDBClusterEndpoint_endpointType :: Lens' CreateDBClusterEndpoint Text

The type of the endpoint. One of: READER, WRITER, ANY.

Destructuring the Response

data DBClusterEndpoint

This data type represents the information you need to connect to an Amazon Aurora DB cluster. This data type is used as a response element in the following actions:

CreateDBClusterEndpoint

DescribeDBClusterEndpoints

ModifyDBClusterEndpoint

DeleteDBClusterEndpoint

For the data structure that represents Amazon RDS DB instance endpoints, see Endpoint.

See: newDBClusterEndpoint smart constructor.

Constructors

DBClusterEndpoint'

Fields

status :: Maybe Text

The current status of the endpoint. One of: creating, available, deleting, inactive, modifying. The inactive state applies to an endpoint that can't be used for a certain kind of cluster, such as a writer endpoint for a read-only secondary cluster in a global database.

dbClusterIdentifier :: Maybe Text

The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

dbClusterEndpointArn :: Maybe Text

The Amazon Resource Name (ARN) for the endpoint.

customEndpointType :: Maybe Text

The type associated with a custom endpoint. One of: READER, WRITER, ANY.

staticMembers :: Maybe [Text]

List of DB instance identifiers that are part of the custom endpoint group.

endpointType :: Maybe Text

The type of the endpoint. One of: READER, WRITER, CUSTOM.

dbClusterEndpointIdentifier :: Maybe Text

The identifier associated with the endpoint. This parameter is stored as a lowercase string.

endpoint :: Maybe Text

The DNS address of the endpoint.

dbClusterEndpointResourceIdentifier :: Maybe Text

A unique system-generated identifier for an endpoint. It remains the same for the whole life of the endpoint.

excludedMembers :: Maybe [Text]

List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

newDBClusterEndpoint :: DBClusterEndpoint

Create a value of DBClusterEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DBClusterEndpoint', dbClusterEndpoint_status - The current status of the endpoint. One of: creating, available, deleting, inactive, modifying. The inactive state applies to an endpoint that can't be used for a certain kind of cluster, such as a writer endpoint for a read-only secondary cluster in a global database.

$sel:dbClusterIdentifier:DBClusterEndpoint', dbClusterEndpoint_dbClusterIdentifier - The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

$sel:dbClusterEndpointArn:DBClusterEndpoint', dbClusterEndpoint_dbClusterEndpointArn - The Amazon Resource Name (ARN) for the endpoint.

$sel:customEndpointType:DBClusterEndpoint', dbClusterEndpoint_customEndpointType - The type associated with a custom endpoint. One of: READER, WRITER, ANY.

$sel:staticMembers:DBClusterEndpoint', dbClusterEndpoint_staticMembers - List of DB instance identifiers that are part of the custom endpoint group.

$sel:endpointType:DBClusterEndpoint', dbClusterEndpoint_endpointType - The type of the endpoint. One of: READER, WRITER, CUSTOM.

$sel:dbClusterEndpointIdentifier:DBClusterEndpoint', dbClusterEndpoint_dbClusterEndpointIdentifier - The identifier associated with the endpoint. This parameter is stored as a lowercase string.

$sel:endpoint:DBClusterEndpoint', dbClusterEndpoint_endpoint - The DNS address of the endpoint.

$sel:dbClusterEndpointResourceIdentifier:DBClusterEndpoint', dbClusterEndpoint_dbClusterEndpointResourceIdentifier - A unique system-generated identifier for an endpoint. It remains the same for the whole life of the endpoint.

$sel:excludedMembers:DBClusterEndpoint', dbClusterEndpoint_excludedMembers - List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

Response Lenses

dbClusterEndpoint_status :: Lens' DBClusterEndpoint (Maybe Text)

The current status of the endpoint. One of: creating, available, deleting, inactive, modifying. The inactive state applies to an endpoint that can't be used for a certain kind of cluster, such as a writer endpoint for a read-only secondary cluster in a global database.

dbClusterEndpoint_dbClusterIdentifier :: Lens' DBClusterEndpoint (Maybe Text)

The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

dbClusterEndpoint_dbClusterEndpointArn :: Lens' DBClusterEndpoint (Maybe Text)

The Amazon Resource Name (ARN) for the endpoint.

dbClusterEndpoint_customEndpointType :: Lens' DBClusterEndpoint (Maybe Text)

The type associated with a custom endpoint. One of: READER, WRITER, ANY.

dbClusterEndpoint_staticMembers :: Lens' DBClusterEndpoint (Maybe [Text])

List of DB instance identifiers that are part of the custom endpoint group.

dbClusterEndpoint_endpointType :: Lens' DBClusterEndpoint (Maybe Text)

The type of the endpoint. One of: READER, WRITER, CUSTOM.

dbClusterEndpoint_dbClusterEndpointIdentifier :: Lens' DBClusterEndpoint (Maybe Text)

The identifier associated with the endpoint. This parameter is stored as a lowercase string.

dbClusterEndpoint_endpoint :: Lens' DBClusterEndpoint (Maybe Text)

The DNS address of the endpoint.

dbClusterEndpoint_dbClusterEndpointResourceIdentifier :: Lens' DBClusterEndpoint (Maybe Text)

A unique system-generated identifier for an endpoint. It remains the same for the whole life of the endpoint.

dbClusterEndpoint_excludedMembers :: Lens' DBClusterEndpoint (Maybe [Text])

List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.