Amazonka.RDS.CreateOptionGroup

Description

Creates a new option group. You can create up to 20 option groups.

Creating a Request

data CreateOptionGroup

See: newCreateOptionGroup smart constructor.

Constructors

CreateOptionGroup'

Fields

tags :: Maybe [Tag]

Tags to assign to the option group.

optionGroupName :: Text

Specifies the name of the option group to be created.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: myoptiongroup

engineName :: Text

Specifies the name of the engine that this option group should be associated with.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

majorEngineVersion :: Text

Specifies the major version of the engine that this option group should be associated with.

optionGroupDescription :: Text

The description of the option group.

newCreateOptionGroup

Arguments

:: Text
$sel:optionGroupName:CreateOptionGroup'

-> Text
$sel:engineName:CreateOptionGroup'

-> Text
$sel:majorEngineVersion:CreateOptionGroup'

-> Text
$sel:optionGroupDescription:CreateOptionGroup'

-> CreateOptionGroup

Create a value of CreateOptionGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateOptionGroup', createOptionGroup_tags - Tags to assign to the option group.

$sel:optionGroupName:CreateOptionGroup', createOptionGroup_optionGroupName - Specifies the name of the option group to be created.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: myoptiongroup

$sel:engineName:CreateOptionGroup', createOptionGroup_engineName - Specifies the name of the engine that this option group should be associated with.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

$sel:majorEngineVersion:CreateOptionGroup', createOptionGroup_majorEngineVersion - Specifies the major version of the engine that this option group should be associated with.

$sel:optionGroupDescription:CreateOptionGroup', createOptionGroup_optionGroupDescription - The description of the option group.

Request Lenses

createOptionGroup_tags :: Lens' CreateOptionGroup (Maybe [Tag])

Tags to assign to the option group.

createOptionGroup_optionGroupName :: Lens' CreateOptionGroup Text

Specifies the name of the option group to be created.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: myoptiongroup

createOptionGroup_engineName :: Lens' CreateOptionGroup Text

Specifies the name of the engine that this option group should be associated with.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

createOptionGroup_majorEngineVersion :: Lens' CreateOptionGroup Text

Specifies the major version of the engine that this option group should be associated with.

createOptionGroup_optionGroupDescription :: Lens' CreateOptionGroup Text

The description of the option group.

Destructuring the Response

data CreateOptionGroupResponse

See: newCreateOptionGroupResponse smart constructor.

Constructors

CreateOptionGroupResponse'

Fields

optionGroup :: Maybe OptionGroup

httpStatus :: Int

The response's http status code.

newCreateOptionGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateOptionGroupResponse'

-> CreateOptionGroupResponse

Create a value of CreateOptionGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionGroup:CreateOptionGroupResponse', createOptionGroupResponse_optionGroup - Undocumented member.

$sel:httpStatus:CreateOptionGroupResponse', createOptionGroupResponse_httpStatus - The response's http status code.

Response Lenses

createOptionGroupResponse_optionGroup :: Lens' CreateOptionGroupResponse (Maybe OptionGroup)

Undocumented member.

createOptionGroupResponse_httpStatus :: Lens' CreateOptionGroupResponse Int

The response's http status code.