Amazonka.RDS.DescribeDBClusterParameters

Description

Returns the detailed parameter list for a particular DB cluster parameter group.

For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This action only applies to Aurora DB clusters.

This operation returns paginated results.

Creating a Request

data DescribeDBClusterParameters

See: newDescribeDBClusterParameters smart constructor.

Constructors

DescribeDBClusterParameters'

Fields

filters :: Maybe [Filter]

This parameter isn't currently supported.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeDBClusterParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

source :: Maybe Text

A value that indicates to return only parameters for a specific source. Parameter sources can be engine, service, or customer.

dbClusterParameterGroupName :: Text

The name of a specific DB cluster parameter group to return parameter details for.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

newDescribeDBClusterParameters

Arguments

:: Text
$sel:dbClusterParameterGroupName:DescribeDBClusterParameters'

-> DescribeDBClusterParameters

Create a value of DescribeDBClusterParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeDBClusterParameters', describeDBClusterParameters_filters - This parameter isn't currently supported.

$sel:marker:DescribeDBClusterParameters', describeDBClusterParameters_marker - An optional pagination token provided by a previous DescribeDBClusterParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeDBClusterParameters', describeDBClusterParameters_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:source:DescribeDBClusterParameters', describeDBClusterParameters_source - A value that indicates to return only parameters for a specific source. Parameter sources can be engine, service, or customer.

$sel:dbClusterParameterGroupName:DescribeDBClusterParameters', describeDBClusterParameters_dbClusterParameterGroupName - The name of a specific DB cluster parameter group to return parameter details for.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

Request Lenses

describeDBClusterParameters_filters :: Lens' DescribeDBClusterParameters (Maybe [Filter])

This parameter isn't currently supported.

describeDBClusterParameters_marker :: Lens' DescribeDBClusterParameters (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBClusterParameters_maxRecords :: Lens' DescribeDBClusterParameters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describeDBClusterParameters_source :: Lens' DescribeDBClusterParameters (Maybe Text)

A value that indicates to return only parameters for a specific source. Parameter sources can be engine, service, or customer.

describeDBClusterParameters_dbClusterParameterGroupName :: Lens' DescribeDBClusterParameters Text

The name of a specific DB cluster parameter group to return parameter details for.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

Destructuring the Response

data DescribeDBClusterParametersResponse

Provides details about a DB cluster parameter group including the parameters in the DB cluster parameter group.

See: newDescribeDBClusterParametersResponse smart constructor.

Constructors

DescribeDBClusterParametersResponse'

Fields

marker :: Maybe Text

An optional pagination token provided by a previous DescribeDBClusterParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

parameters :: Maybe [Parameter]

Provides a list of parameters for the DB cluster parameter group.

httpStatus :: Int

The response's http status code.

newDescribeDBClusterParametersResponse

Arguments

:: Int
$sel:httpStatus:DescribeDBClusterParametersResponse'

-> DescribeDBClusterParametersResponse

Create a value of DescribeDBClusterParametersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeDBClusterParameters', describeDBClusterParametersResponse_marker - An optional pagination token provided by a previous DescribeDBClusterParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

$sel:parameters:DescribeDBClusterParametersResponse', describeDBClusterParametersResponse_parameters - Provides a list of parameters for the DB cluster parameter group.

$sel:httpStatus:DescribeDBClusterParametersResponse', describeDBClusterParametersResponse_httpStatus - The response's http status code.

Response Lenses

describeDBClusterParametersResponse_marker :: Lens' DescribeDBClusterParametersResponse (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterParameters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

describeDBClusterParametersResponse_parameters :: Lens' DescribeDBClusterParametersResponse (Maybe [Parameter])

Provides a list of parameters for the DB cluster parameter group.

describeDBClusterParametersResponse_httpStatus :: Lens' DescribeDBClusterParametersResponse Int

The response's http status code.