Amazonka.RDS.DescribeDBSecurityGroups

Description

Returns a list of DBSecurityGroup descriptions. If a DBSecurityGroupName is specified, the list will contain only the descriptions of the specified DB security group.

This operation returns paginated results.

Creating a Request

data DescribeDBSecurityGroups

See: newDescribeDBSecurityGroups smart constructor.

Constructors

DescribeDBSecurityGroups'

Fields

filters :: Maybe [Filter]

This parameter isn't currently supported.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeDBSecurityGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

dbSecurityGroupName :: Maybe Text

The name of the DB security group to return details for.

newDescribeDBSecurityGroups :: DescribeDBSecurityGroups

Create a value of DescribeDBSecurityGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeDBSecurityGroups', describeDBSecurityGroups_filters - This parameter isn't currently supported.

$sel:marker:DescribeDBSecurityGroups', describeDBSecurityGroups_marker - An optional pagination token provided by a previous DescribeDBSecurityGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeDBSecurityGroups', describeDBSecurityGroups_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:dbSecurityGroupName:DescribeDBSecurityGroups', describeDBSecurityGroups_dbSecurityGroupName - The name of the DB security group to return details for.

Request Lenses

describeDBSecurityGroups_filters :: Lens' DescribeDBSecurityGroups (Maybe [Filter])

This parameter isn't currently supported.

describeDBSecurityGroups_marker :: Lens' DescribeDBSecurityGroups (Maybe Text)

An optional pagination token provided by a previous DescribeDBSecurityGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBSecurityGroups_maxRecords :: Lens' DescribeDBSecurityGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describeDBSecurityGroups_dbSecurityGroupName :: Lens' DescribeDBSecurityGroups (Maybe Text)

The name of the DB security group to return details for.

Destructuring the Response

data DescribeDBSecurityGroupsResponse

Contains the result of a successful invocation of the DescribeDBSecurityGroups action.

See: newDescribeDBSecurityGroupsResponse smart constructor.

Constructors

DescribeDBSecurityGroupsResponse'

Fields

dbSecurityGroups :: Maybe [DBSecurityGroup]

A list of DBSecurityGroup instances.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeDBSecurityGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDBSecurityGroupsResponse'

-> DescribeDBSecurityGroupsResponse

Create a value of DescribeDBSecurityGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSecurityGroups:DescribeDBSecurityGroupsResponse', describeDBSecurityGroupsResponse_dbSecurityGroups - A list of DBSecurityGroup instances.

$sel:marker:DescribeDBSecurityGroups', describeDBSecurityGroupsResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeDBSecurityGroupsResponse', describeDBSecurityGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeDBSecurityGroupsResponse_dbSecurityGroups :: Lens' DescribeDBSecurityGroupsResponse (Maybe [DBSecurityGroup])

A list of DBSecurityGroup instances.

describeDBSecurityGroupsResponse_marker :: Lens' DescribeDBSecurityGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBSecurityGroupsResponse_httpStatus :: Lens' DescribeDBSecurityGroupsResponse Int

The response's http status code.