Amazonka.RDS.DescribeOptionGroups

Description

Describes the available option groups.

This operation returns paginated results.

Creating a Request

data DescribeOptionGroups

See: newDescribeOptionGroups smart constructor.

Constructors

DescribeOptionGroups'

Fields

filters :: Maybe [Filter]

This parameter isn't currently supported.

engineName :: Maybe Text

Filters the list of option groups to only include groups associated with a specific database engine.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

majorEngineVersion :: Maybe Text

Filters the list of option groups to only include groups associated with a specific database engine version. If specified, then EngineName must also be specified.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeOptionGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

optionGroupName :: Maybe Text

The name of the option group to describe. Can't be supplied together with EngineName or MajorEngineVersion.

newDescribeOptionGroups :: DescribeOptionGroups

Create a value of DescribeOptionGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeOptionGroups', describeOptionGroups_filters - This parameter isn't currently supported.

$sel:engineName:DescribeOptionGroups', describeOptionGroups_engineName - Filters the list of option groups to only include groups associated with a specific database engine.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

$sel:majorEngineVersion:DescribeOptionGroups', describeOptionGroups_majorEngineVersion - Filters the list of option groups to only include groups associated with a specific database engine version. If specified, then EngineName must also be specified.

$sel:marker:DescribeOptionGroups', describeOptionGroups_marker - An optional pagination token provided by a previous DescribeOptionGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeOptionGroups', describeOptionGroups_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:optionGroupName:DescribeOptionGroups', describeOptionGroups_optionGroupName - The name of the option group to describe. Can't be supplied together with EngineName or MajorEngineVersion.

Request Lenses

describeOptionGroups_filters :: Lens' DescribeOptionGroups (Maybe [Filter])

This parameter isn't currently supported.

describeOptionGroups_engineName :: Lens' DescribeOptionGroups (Maybe Text)

Filters the list of option groups to only include groups associated with a specific database engine.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

describeOptionGroups_majorEngineVersion :: Lens' DescribeOptionGroups (Maybe Text)

Filters the list of option groups to only include groups associated with a specific database engine version. If specified, then EngineName must also be specified.

describeOptionGroups_marker :: Lens' DescribeOptionGroups (Maybe Text)

An optional pagination token provided by a previous DescribeOptionGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeOptionGroups_maxRecords :: Lens' DescribeOptionGroups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describeOptionGroups_optionGroupName :: Lens' DescribeOptionGroups (Maybe Text)

The name of the option group to describe. Can't be supplied together with EngineName or MajorEngineVersion.

Destructuring the Response

data DescribeOptionGroupsResponse

List of option groups.

See: newDescribeOptionGroupsResponse smart constructor.

Constructors

DescribeOptionGroupsResponse'

Fields

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

optionGroupsList :: Maybe [OptionGroup]

List of option groups.

httpStatus :: Int

The response's http status code.

newDescribeOptionGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeOptionGroupsResponse'

-> DescribeOptionGroupsResponse

Create a value of DescribeOptionGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeOptionGroups', describeOptionGroupsResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:optionGroupsList:DescribeOptionGroupsResponse', describeOptionGroupsResponse_optionGroupsList - List of option groups.

$sel:httpStatus:DescribeOptionGroupsResponse', describeOptionGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeOptionGroupsResponse_marker :: Lens' DescribeOptionGroupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeOptionGroupsResponse_optionGroupsList :: Lens' DescribeOptionGroupsResponse (Maybe [OptionGroup])

List of option groups.

describeOptionGroupsResponse_httpStatus :: Lens' DescribeOptionGroupsResponse Int

The response's http status code.