Amazonka.RDS.DescribePendingMaintenanceActions

Description

Returns a list of resources (for example, DB instances) that have at least one pending maintenance action.

This operation returns paginated results.

Creating a Request

data DescribePendingMaintenanceActions

See: newDescribePendingMaintenanceActions smart constructor.

Constructors

DescribePendingMaintenanceActions'

Fields

filters :: Maybe [Filter]

A filter that specifies one or more resources to return pending maintenance actions for.

Supported filters:

db-cluster-id - Accepts DB cluster identifiers and DB cluster Amazon Resource Names (ARNs). The results list will only include pending maintenance actions for the DB clusters identified by these ARNs.

db-instance-id - Accepts DB instance identifiers and DB instance ARNs. The results list will only include pending maintenance actions for the DB instances identified by these ARNs.

marker :: Maybe Text

An optional pagination token provided by a previous DescribePendingMaintenanceActions request. If this parameter is specified, the response includes only records beyond the marker, up to a number of records specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

resourceIdentifier :: Maybe Text

The ARN of a resource to return pending maintenance actions for.

newDescribePendingMaintenanceActions :: DescribePendingMaintenanceActions

Create a value of DescribePendingMaintenanceActions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribePendingMaintenanceActions', describePendingMaintenanceActions_filters - A filter that specifies one or more resources to return pending maintenance actions for.

Supported filters:

db-cluster-id - Accepts DB cluster identifiers and DB cluster Amazon Resource Names (ARNs). The results list will only include pending maintenance actions for the DB clusters identified by these ARNs.

db-instance-id - Accepts DB instance identifiers and DB instance ARNs. The results list will only include pending maintenance actions for the DB instances identified by these ARNs.

$sel:marker:DescribePendingMaintenanceActions', describePendingMaintenanceActions_marker - An optional pagination token provided by a previous DescribePendingMaintenanceActions request. If this parameter is specified, the response includes only records beyond the marker, up to a number of records specified by MaxRecords.

$sel:maxRecords:DescribePendingMaintenanceActions', describePendingMaintenanceActions_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:resourceIdentifier:DescribePendingMaintenanceActions', describePendingMaintenanceActions_resourceIdentifier - The ARN of a resource to return pending maintenance actions for.

Request Lenses

describePendingMaintenanceActions_filters :: Lens' DescribePendingMaintenanceActions (Maybe [Filter])

A filter that specifies one or more resources to return pending maintenance actions for.

Supported filters:

db-cluster-id - Accepts DB cluster identifiers and DB cluster Amazon Resource Names (ARNs). The results list will only include pending maintenance actions for the DB clusters identified by these ARNs.

db-instance-id - Accepts DB instance identifiers and DB instance ARNs. The results list will only include pending maintenance actions for the DB instances identified by these ARNs.

describePendingMaintenanceActions_marker :: Lens' DescribePendingMaintenanceActions (Maybe Text)

An optional pagination token provided by a previous DescribePendingMaintenanceActions request. If this parameter is specified, the response includes only records beyond the marker, up to a number of records specified by MaxRecords.

describePendingMaintenanceActions_maxRecords :: Lens' DescribePendingMaintenanceActions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describePendingMaintenanceActions_resourceIdentifier :: Lens' DescribePendingMaintenanceActions (Maybe Text)

The ARN of a resource to return pending maintenance actions for.

Destructuring the Response

data DescribePendingMaintenanceActionsResponse

Data returned from the DescribePendingMaintenanceActions action.

See: newDescribePendingMaintenanceActionsResponse smart constructor.

Constructors

DescribePendingMaintenanceActionsResponse'

Fields

pendingMaintenanceActions :: Maybe [ResourcePendingMaintenanceActions]

A list of the pending maintenance actions for the resource.

marker :: Maybe Text

An optional pagination token provided by a previous DescribePendingMaintenanceActions request. If this parameter is specified, the response includes only records beyond the marker, up to a number of records specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribePendingMaintenanceActionsResponse

Arguments

:: Int
$sel:httpStatus:DescribePendingMaintenanceActionsResponse'

-> DescribePendingMaintenanceActionsResponse

Create a value of DescribePendingMaintenanceActionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pendingMaintenanceActions:DescribePendingMaintenanceActionsResponse', describePendingMaintenanceActionsResponse_pendingMaintenanceActions - A list of the pending maintenance actions for the resource.

$sel:marker:DescribePendingMaintenanceActions', describePendingMaintenanceActionsResponse_marker - An optional pagination token provided by a previous DescribePendingMaintenanceActions request. If this parameter is specified, the response includes only records beyond the marker, up to a number of records specified by MaxRecords.

$sel:httpStatus:DescribePendingMaintenanceActionsResponse', describePendingMaintenanceActionsResponse_httpStatus - The response's http status code.

Response Lenses

describePendingMaintenanceActionsResponse_pendingMaintenanceActions :: Lens' DescribePendingMaintenanceActionsResponse (Maybe [ResourcePendingMaintenanceActions])

A list of the pending maintenance actions for the resource.

describePendingMaintenanceActionsResponse_marker :: Lens' DescribePendingMaintenanceActionsResponse (Maybe Text)

An optional pagination token provided by a previous DescribePendingMaintenanceActions request. If this parameter is specified, the response includes only records beyond the marker, up to a number of records specified by MaxRecords.

describePendingMaintenanceActionsResponse_httpStatus :: Lens' DescribePendingMaintenanceActionsResponse Int

The response's http status code.