Amazonka.RDS.DescribeValidDBInstanceModifications

Description

You can call DescribeValidDBInstanceModifications to learn what modifications you can make to your DB instance. You can use this information when you call ModifyDBInstance.

Creating a Request

data DescribeValidDBInstanceModifications

See: newDescribeValidDBInstanceModifications smart constructor.

Constructors

DescribeValidDBInstanceModifications'

Fields

dbInstanceIdentifier :: Text

The customer identifier or the ARN of your DB instance.

newDescribeValidDBInstanceModifications

Arguments

:: Text
$sel:dbInstanceIdentifier:DescribeValidDBInstanceModifications'

-> DescribeValidDBInstanceModifications

Create a value of DescribeValidDBInstanceModifications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstanceIdentifier:DescribeValidDBInstanceModifications', describeValidDBInstanceModifications_dbInstanceIdentifier - The customer identifier or the ARN of your DB instance.

Request Lenses

describeValidDBInstanceModifications_dbInstanceIdentifier :: Lens' DescribeValidDBInstanceModifications Text

The customer identifier or the ARN of your DB instance.

Destructuring the Response

data DescribeValidDBInstanceModificationsResponse

See: newDescribeValidDBInstanceModificationsResponse smart constructor.

Constructors

DescribeValidDBInstanceModificationsResponse'

Fields

validDBInstanceModificationsMessage :: Maybe ValidDBInstanceModificationsMessage

httpStatus :: Int

The response's http status code.

newDescribeValidDBInstanceModificationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeValidDBInstanceModificationsResponse'

-> DescribeValidDBInstanceModificationsResponse

Create a value of DescribeValidDBInstanceModificationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validDBInstanceModificationsMessage:DescribeValidDBInstanceModificationsResponse', describeValidDBInstanceModificationsResponse_validDBInstanceModificationsMessage - Undocumented member.

$sel:httpStatus:DescribeValidDBInstanceModificationsResponse', describeValidDBInstanceModificationsResponse_httpStatus - The response's http status code.

Response Lenses

describeValidDBInstanceModificationsResponse_validDBInstanceModificationsMessage :: Lens' DescribeValidDBInstanceModificationsResponse (Maybe ValidDBInstanceModificationsMessage)

Undocumented member.

describeValidDBInstanceModificationsResponse_httpStatus :: Lens' DescribeValidDBInstanceModificationsResponse Int

The response's http status code.