Amazonka.RDS.DownloadDBLogFilePortion

Description

Downloads all or a portion of the specified log file, up to 1 MB in size.

This operation returns paginated results.

Creating a Request

data DownloadDBLogFilePortion

See: newDownloadDBLogFilePortion smart constructor.

Constructors

DownloadDBLogFilePortion'

Fields

numberOfLines :: Maybe Int

The number of lines to download. If the number of lines specified results in a file over 1 MB in size, the file is truncated at 1 MB in size.

If the NumberOfLines parameter is specified, then the block of lines returned can be from the beginning or the end of the log file, depending on the value of the Marker parameter.

If neither Marker or NumberOfLines are specified, the entire log file is returned up to a maximum of 10000 lines, starting with the most recent log entries first.

If NumberOfLines is specified and Marker isn't specified, then the most recent lines from the end of the log file are returned.

If Marker is specified as "0", then the specified number of lines from the beginning of the log file are returned.

You can download the log file in blocks of lines by specifying the size of the block using the NumberOfLines parameter, and by specifying a value of "0" for the Marker parameter in your first request. Include the Marker value returned in the response as the Marker value for the next request, continuing until the AdditionalDataPending response element returns false.

marker :: Maybe Text

The pagination token provided in the previous request or "0". If the Marker parameter is specified the response includes only records beyond the marker until the end of the file or up to NumberOfLines.

dbInstanceIdentifier :: Text

The customer-assigned name of the DB instance that contains the log files you want to list.

Constraints:

Must match the identifier of an existing DBInstance.

logFileName :: Text

The name of the log file to be downloaded.

newDownloadDBLogFilePortion

Arguments

:: Text
$sel:dbInstanceIdentifier:DownloadDBLogFilePortion'

-> Text
$sel:logFileName:DownloadDBLogFilePortion'

-> DownloadDBLogFilePortion

Create a value of DownloadDBLogFilePortion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numberOfLines:DownloadDBLogFilePortion', downloadDBLogFilePortion_numberOfLines - The number of lines to download. If the number of lines specified results in a file over 1 MB in size, the file is truncated at 1 MB in size.

If the NumberOfLines parameter is specified, then the block of lines returned can be from the beginning or the end of the log file, depending on the value of the Marker parameter.

If neither Marker or NumberOfLines are specified, the entire log file is returned up to a maximum of 10000 lines, starting with the most recent log entries first.

If NumberOfLines is specified and Marker isn't specified, then the most recent lines from the end of the log file are returned.

If Marker is specified as "0", then the specified number of lines from the beginning of the log file are returned.

You can download the log file in blocks of lines by specifying the size of the block using the NumberOfLines parameter, and by specifying a value of "0" for the Marker parameter in your first request. Include the Marker value returned in the response as the Marker value for the next request, continuing until the AdditionalDataPending response element returns false.

$sel:marker:DownloadDBLogFilePortion', downloadDBLogFilePortion_marker - The pagination token provided in the previous request or "0". If the Marker parameter is specified the response includes only records beyond the marker until the end of the file or up to NumberOfLines.

$sel:dbInstanceIdentifier:DownloadDBLogFilePortion', downloadDBLogFilePortion_dbInstanceIdentifier - The customer-assigned name of the DB instance that contains the log files you want to list.

Constraints:

Must match the identifier of an existing DBInstance.

$sel:logFileName:DownloadDBLogFilePortion', downloadDBLogFilePortion_logFileName - The name of the log file to be downloaded.

Request Lenses

downloadDBLogFilePortion_numberOfLines :: Lens' DownloadDBLogFilePortion (Maybe Int)

The number of lines to download. If the number of lines specified results in a file over 1 MB in size, the file is truncated at 1 MB in size.

If the NumberOfLines parameter is specified, then the block of lines returned can be from the beginning or the end of the log file, depending on the value of the Marker parameter.

If neither Marker or NumberOfLines are specified, the entire log file is returned up to a maximum of 10000 lines, starting with the most recent log entries first.

If NumberOfLines is specified and Marker isn't specified, then the most recent lines from the end of the log file are returned.

If Marker is specified as "0", then the specified number of lines from the beginning of the log file are returned.

You can download the log file in blocks of lines by specifying the size of the block using the NumberOfLines parameter, and by specifying a value of "0" for the Marker parameter in your first request. Include the Marker value returned in the response as the Marker value for the next request, continuing until the AdditionalDataPending response element returns false.

downloadDBLogFilePortion_marker :: Lens' DownloadDBLogFilePortion (Maybe Text)

The pagination token provided in the previous request or "0". If the Marker parameter is specified the response includes only records beyond the marker until the end of the file or up to NumberOfLines.

downloadDBLogFilePortion_dbInstanceIdentifier :: Lens' DownloadDBLogFilePortion Text

The customer-assigned name of the DB instance that contains the log files you want to list.

Constraints:

Must match the identifier of an existing DBInstance.

downloadDBLogFilePortion_logFileName :: Lens' DownloadDBLogFilePortion Text

The name of the log file to be downloaded.

Destructuring the Response

data DownloadDBLogFilePortionResponse

This data type is used as a response element to DownloadDBLogFilePortion.

See: newDownloadDBLogFilePortionResponse smart constructor.

Constructors

DownloadDBLogFilePortionResponse'

Fields

logFileData :: Maybe Text

Entries from the specified log file.

additionalDataPending :: Maybe Bool

Boolean value that if true, indicates there is more data to be downloaded.

marker :: Maybe Text

A pagination token that can be used in a later DownloadDBLogFilePortion request.

httpStatus :: Int

The response's http status code.

newDownloadDBLogFilePortionResponse

Arguments

:: Int
$sel:httpStatus:DownloadDBLogFilePortionResponse'

-> DownloadDBLogFilePortionResponse

Create a value of DownloadDBLogFilePortionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logFileData:DownloadDBLogFilePortionResponse', downloadDBLogFilePortionResponse_logFileData - Entries from the specified log file.

$sel:additionalDataPending:DownloadDBLogFilePortionResponse', downloadDBLogFilePortionResponse_additionalDataPending - Boolean value that if true, indicates there is more data to be downloaded.

$sel:marker:DownloadDBLogFilePortion', downloadDBLogFilePortionResponse_marker - A pagination token that can be used in a later DownloadDBLogFilePortion request.

$sel:httpStatus:DownloadDBLogFilePortionResponse', downloadDBLogFilePortionResponse_httpStatus - The response's http status code.

Response Lenses

downloadDBLogFilePortionResponse_logFileData :: Lens' DownloadDBLogFilePortionResponse (Maybe Text)

Entries from the specified log file.

downloadDBLogFilePortionResponse_additionalDataPending :: Lens' DownloadDBLogFilePortionResponse (Maybe Bool)

Boolean value that if true, indicates there is more data to be downloaded.

downloadDBLogFilePortionResponse_marker :: Lens' DownloadDBLogFilePortionResponse (Maybe Text)

A pagination token that can be used in a later DownloadDBLogFilePortion request.

downloadDBLogFilePortionResponse_httpStatus :: Lens' DownloadDBLogFilePortionResponse Int

The response's http status code.