Amazonka.RDS.ModifyCurrentDBClusterCapacity

Description

Set the capacity of an Aurora Serverless DB cluster to a specific value.

Aurora Serverless scales seamlessly based on the workload on the DB cluster. In some cases, the capacity might not scale fast enough to meet a sudden change in workload, such as a large number of new transactions. Call ModifyCurrentDBClusterCapacity to set the capacity explicitly.

After this call sets the DB cluster capacity, Aurora Serverless can automatically scale the DB cluster based on the cooldown period for scaling up and the cooldown period for scaling down.

For more information about Aurora Serverless, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

If you call ModifyCurrentDBClusterCapacity with the default TimeoutAction, connections that prevent Aurora Serverless from finding a scaling point might be dropped. For more information about scaling points, see Autoscaling for Aurora Serverless in the Amazon Aurora User Guide.

This action only applies to Aurora Serverless DB clusters.

Creating a Request

data ModifyCurrentDBClusterCapacity

See: newModifyCurrentDBClusterCapacity smart constructor.

Constructors

ModifyCurrentDBClusterCapacity'

Fields

timeoutAction :: Maybe Text

The action to take when the timeout is reached, either ForceApplyCapacityChange or RollbackCapacityChange.

ForceApplyCapacityChange, the default, sets the capacity to the specified value as soon as possible.

RollbackCapacityChange ignores the capacity change if a scaling point isn't found in the timeout period.

capacity :: Maybe Int

The DB cluster capacity.

When you change the capacity of a paused Aurora Serverless DB cluster, it automatically resumes.

Constraints:

For Aurora MySQL, valid capacity values are 1, 2, 4, 8, 16, 32, 64, 128, and 256.

For Aurora PostgreSQL, valid capacity values are 2, 4, 8, 16, 32, 64, 192, and 384.

secondsBeforeTimeout :: Maybe Int

The amount of time, in seconds, that Aurora Serverless tries to find a scaling point to perform seamless scaling before enforcing the timeout action. The default is 300.

Specify a value between 10 and 600 seconds.

dbClusterIdentifier :: Text

The DB cluster identifier for the cluster being modified. This parameter isn't case-sensitive.

Constraints:

Must match the identifier of an existing DB cluster.

newModifyCurrentDBClusterCapacity

Arguments

:: Text
$sel:dbClusterIdentifier:ModifyCurrentDBClusterCapacity'

-> ModifyCurrentDBClusterCapacity

Create a value of ModifyCurrentDBClusterCapacity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeoutAction:ModifyCurrentDBClusterCapacity', modifyCurrentDBClusterCapacity_timeoutAction - The action to take when the timeout is reached, either ForceApplyCapacityChange or RollbackCapacityChange.

ForceApplyCapacityChange, the default, sets the capacity to the specified value as soon as possible.

RollbackCapacityChange ignores the capacity change if a scaling point isn't found in the timeout period.

$sel:capacity:ModifyCurrentDBClusterCapacity', modifyCurrentDBClusterCapacity_capacity - The DB cluster capacity.

When you change the capacity of a paused Aurora Serverless DB cluster, it automatically resumes.

Constraints:

For Aurora MySQL, valid capacity values are 1, 2, 4, 8, 16, 32, 64, 128, and 256.

For Aurora PostgreSQL, valid capacity values are 2, 4, 8, 16, 32, 64, 192, and 384.

$sel:secondsBeforeTimeout:ModifyCurrentDBClusterCapacity', modifyCurrentDBClusterCapacity_secondsBeforeTimeout - The amount of time, in seconds, that Aurora Serverless tries to find a scaling point to perform seamless scaling before enforcing the timeout action. The default is 300.

Specify a value between 10 and 600 seconds.

$sel:dbClusterIdentifier:ModifyCurrentDBClusterCapacity', modifyCurrentDBClusterCapacity_dbClusterIdentifier - The DB cluster identifier for the cluster being modified. This parameter isn't case-sensitive.

Constraints:

Must match the identifier of an existing DB cluster.

Request Lenses

modifyCurrentDBClusterCapacity_timeoutAction :: Lens' ModifyCurrentDBClusterCapacity (Maybe Text)

The action to take when the timeout is reached, either ForceApplyCapacityChange or RollbackCapacityChange.

ForceApplyCapacityChange, the default, sets the capacity to the specified value as soon as possible.

RollbackCapacityChange ignores the capacity change if a scaling point isn't found in the timeout period.

modifyCurrentDBClusterCapacity_capacity :: Lens' ModifyCurrentDBClusterCapacity (Maybe Int)

The DB cluster capacity.

When you change the capacity of a paused Aurora Serverless DB cluster, it automatically resumes.

Constraints:

For Aurora MySQL, valid capacity values are 1, 2, 4, 8, 16, 32, 64, 128, and 256.

For Aurora PostgreSQL, valid capacity values are 2, 4, 8, 16, 32, 64, 192, and 384.

modifyCurrentDBClusterCapacity_secondsBeforeTimeout :: Lens' ModifyCurrentDBClusterCapacity (Maybe Int)

The amount of time, in seconds, that Aurora Serverless tries to find a scaling point to perform seamless scaling before enforcing the timeout action. The default is 300.

Specify a value between 10 and 600 seconds.

modifyCurrentDBClusterCapacity_dbClusterIdentifier :: Lens' ModifyCurrentDBClusterCapacity Text

The DB cluster identifier for the cluster being modified. This parameter isn't case-sensitive.

Constraints:

Must match the identifier of an existing DB cluster.

Destructuring the Response

data ModifyCurrentDBClusterCapacityResponse

See: newModifyCurrentDBClusterCapacityResponse smart constructor.

Constructors

ModifyCurrentDBClusterCapacityResponse'

Fields

dbClusterIdentifier :: Maybe Text

A user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

timeoutAction :: Maybe Text

The timeout action of a call to ModifyCurrentDBClusterCapacity, either ForceApplyCapacityChange or RollbackCapacityChange.

currentCapacity :: Maybe Int

The current capacity of the DB cluster.

pendingCapacity :: Maybe Int

A value that specifies the capacity that the DB cluster scales to next.

secondsBeforeTimeout :: Maybe Int

The number of seconds before a call to ModifyCurrentDBClusterCapacity times out.

httpStatus :: Int

The response's http status code.

newModifyCurrentDBClusterCapacityResponse

Arguments

:: Int
$sel:httpStatus:ModifyCurrentDBClusterCapacityResponse'

-> ModifyCurrentDBClusterCapacityResponse

Create a value of ModifyCurrentDBClusterCapacityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterIdentifier:ModifyCurrentDBClusterCapacity', modifyCurrentDBClusterCapacityResponse_dbClusterIdentifier - A user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

$sel:timeoutAction:ModifyCurrentDBClusterCapacity', modifyCurrentDBClusterCapacityResponse_timeoutAction - The timeout action of a call to ModifyCurrentDBClusterCapacity, either ForceApplyCapacityChange or RollbackCapacityChange.

$sel:currentCapacity:ModifyCurrentDBClusterCapacityResponse', modifyCurrentDBClusterCapacityResponse_currentCapacity - The current capacity of the DB cluster.

$sel:pendingCapacity:ModifyCurrentDBClusterCapacityResponse', modifyCurrentDBClusterCapacityResponse_pendingCapacity - A value that specifies the capacity that the DB cluster scales to next.

$sel:secondsBeforeTimeout:ModifyCurrentDBClusterCapacity', modifyCurrentDBClusterCapacityResponse_secondsBeforeTimeout - The number of seconds before a call to ModifyCurrentDBClusterCapacity times out.

$sel:httpStatus:ModifyCurrentDBClusterCapacityResponse', modifyCurrentDBClusterCapacityResponse_httpStatus - The response's http status code.

Response Lenses

modifyCurrentDBClusterCapacityResponse_dbClusterIdentifier :: Lens' ModifyCurrentDBClusterCapacityResponse (Maybe Text)

A user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

modifyCurrentDBClusterCapacityResponse_timeoutAction :: Lens' ModifyCurrentDBClusterCapacityResponse (Maybe Text)

The timeout action of a call to ModifyCurrentDBClusterCapacity, either ForceApplyCapacityChange or RollbackCapacityChange.

modifyCurrentDBClusterCapacityResponse_currentCapacity :: Lens' ModifyCurrentDBClusterCapacityResponse (Maybe Int)

The current capacity of the DB cluster.

modifyCurrentDBClusterCapacityResponse_pendingCapacity :: Lens' ModifyCurrentDBClusterCapacityResponse (Maybe Int)

A value that specifies the capacity that the DB cluster scales to next.

modifyCurrentDBClusterCapacityResponse_secondsBeforeTimeout :: Lens' ModifyCurrentDBClusterCapacityResponse (Maybe Int)

The number of seconds before a call to ModifyCurrentDBClusterCapacity times out.

modifyCurrentDBClusterCapacityResponse_httpStatus :: Lens' ModifyCurrentDBClusterCapacityResponse Int

The response's http status code.