Amazonka.RDS.RemoveRoleFromDBInstance

Description

Disassociates an Amazon Web Services Identity and Access Management (IAM) role from a DB instance.

Creating a Request

data RemoveRoleFromDBInstance

See: newRemoveRoleFromDBInstance smart constructor.

Constructors

RemoveRoleFromDBInstance'

Fields

dbInstanceIdentifier :: Text

The name of the DB instance to disassociate the IAM role from.

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM role to disassociate from the DB instance, for example, arn:aws:iam::123456789012:role/AccessRole.

featureName :: Text

The name of the feature for the DB instance that the IAM role is to be disassociated from. For the list of supported feature names, see DBEngineVersion.

newRemoveRoleFromDBInstance

Arguments

:: Text
$sel:dbInstanceIdentifier:RemoveRoleFromDBInstance'

-> Text
$sel:roleArn:RemoveRoleFromDBInstance'

-> Text
$sel:featureName:RemoveRoleFromDBInstance'

-> RemoveRoleFromDBInstance

Create a value of RemoveRoleFromDBInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstanceIdentifier:RemoveRoleFromDBInstance', removeRoleFromDBInstance_dbInstanceIdentifier - The name of the DB instance to disassociate the IAM role from.

$sel:roleArn:RemoveRoleFromDBInstance', removeRoleFromDBInstance_roleArn - The Amazon Resource Name (ARN) of the IAM role to disassociate from the DB instance, for example, arn:aws:iam::123456789012:role/AccessRole.

$sel:featureName:RemoveRoleFromDBInstance', removeRoleFromDBInstance_featureName - The name of the feature for the DB instance that the IAM role is to be disassociated from. For the list of supported feature names, see DBEngineVersion.

Request Lenses

removeRoleFromDBInstance_dbInstanceIdentifier :: Lens' RemoveRoleFromDBInstance Text

The name of the DB instance to disassociate the IAM role from.

removeRoleFromDBInstance_roleArn :: Lens' RemoveRoleFromDBInstance Text

The Amazon Resource Name (ARN) of the IAM role to disassociate from the DB instance, for example, arn:aws:iam::123456789012:role/AccessRole.

removeRoleFromDBInstance_featureName :: Lens' RemoveRoleFromDBInstance Text

The name of the feature for the DB instance that the IAM role is to be disassociated from. For the list of supported feature names, see DBEngineVersion.

Destructuring the Response

data RemoveRoleFromDBInstanceResponse

See: newRemoveRoleFromDBInstanceResponse smart constructor.

Constructors

RemoveRoleFromDBInstanceResponse'

newRemoveRoleFromDBInstanceResponse :: RemoveRoleFromDBInstanceResponse

Create a value of RemoveRoleFromDBInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.