Amazonka.RDS.ResetDBParameterGroup

Description

Modifies the parameters of a DB parameter group to the engine/system default value. To reset specific parameters, provide a list of the following: ParameterName and ApplyMethod. To reset the entire DB parameter group, specify the DBParameterGroup name and ResetAllParameters parameters. When resetting the entire group, dynamic parameters are updated immediately and static parameters are set to pending-reboot to take effect on the next DB instance restart or RebootDBInstance request.

Creating a Request

data ResetDBParameterGroup

See: newResetDBParameterGroup smart constructor.

Constructors

ResetDBParameterGroup'

Fields

resetAllParameters :: Maybe Bool

A value that indicates whether to reset all parameters in the DB parameter group to default values. By default, all parameters in the DB parameter group are reset to default values.

parameters :: Maybe [Parameter]

To reset the entire DB parameter group, specify the DBParameterGroup name and ResetAllParameters parameters. To reset specific parameters, provide a list of the following: ParameterName and ApplyMethod. A maximum of 20 parameters can be modified in a single request.

MySQL

Valid Values (for Apply method): immediate | pending-reboot

You can use the immediate value with dynamic parameters only. You can use the pending-reboot value for both dynamic and static parameters, and changes are applied when DB instance reboots.

MariaDB

Valid Values (for Apply method): immediate | pending-reboot

You can use the immediate value with dynamic parameters only. You can use the pending-reboot value for both dynamic and static parameters, and changes are applied when DB instance reboots.

Oracle

Valid Values (for Apply method): pending-reboot

dbParameterGroupName :: Text

The name of the DB parameter group.

Constraints:

Must match the name of an existing DBParameterGroup.

newResetDBParameterGroup

Arguments

:: Text
$sel:dbParameterGroupName:ResetDBParameterGroup'

-> ResetDBParameterGroup

Create a value of ResetDBParameterGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resetAllParameters:ResetDBParameterGroup', resetDBParameterGroup_resetAllParameters - A value that indicates whether to reset all parameters in the DB parameter group to default values. By default, all parameters in the DB parameter group are reset to default values.

$sel:parameters:ResetDBParameterGroup', resetDBParameterGroup_parameters - To reset the entire DB parameter group, specify the DBParameterGroup name and ResetAllParameters parameters. To reset specific parameters, provide a list of the following: ParameterName and ApplyMethod. A maximum of 20 parameters can be modified in a single request.

MySQL

Valid Values (for Apply method): immediate | pending-reboot

You can use the immediate value with dynamic parameters only. You can use the pending-reboot value for both dynamic and static parameters, and changes are applied when DB instance reboots.

MariaDB

Valid Values (for Apply method): immediate | pending-reboot

You can use the immediate value with dynamic parameters only. You can use the pending-reboot value for both dynamic and static parameters, and changes are applied when DB instance reboots.

Oracle

Valid Values (for Apply method): pending-reboot

$sel:dbParameterGroupName:ResetDBParameterGroup', resetDBParameterGroup_dbParameterGroupName - The name of the DB parameter group.

Constraints:

Must match the name of an existing DBParameterGroup.

Request Lenses

resetDBParameterGroup_resetAllParameters :: Lens' ResetDBParameterGroup (Maybe Bool)

A value that indicates whether to reset all parameters in the DB parameter group to default values. By default, all parameters in the DB parameter group are reset to default values.

resetDBParameterGroup_parameters :: Lens' ResetDBParameterGroup (Maybe [Parameter])

To reset the entire DB parameter group, specify the DBParameterGroup name and ResetAllParameters parameters. To reset specific parameters, provide a list of the following: ParameterName and ApplyMethod. A maximum of 20 parameters can be modified in a single request.

MySQL

Valid Values (for Apply method): immediate | pending-reboot

You can use the immediate value with dynamic parameters only. You can use the pending-reboot value for both dynamic and static parameters, and changes are applied when DB instance reboots.

MariaDB

Valid Values (for Apply method): immediate | pending-reboot

You can use the immediate value with dynamic parameters only. You can use the pending-reboot value for both dynamic and static parameters, and changes are applied when DB instance reboots.

Oracle

Valid Values (for Apply method): pending-reboot

resetDBParameterGroup_dbParameterGroupName :: Lens' ResetDBParameterGroup Text

The name of the DB parameter group.

Constraints:

Must match the name of an existing DBParameterGroup.

Destructuring the Response

data DBParameterGroupNameMessage

Contains the result of a successful invocation of the ModifyDBParameterGroup or ResetDBParameterGroup action.

See: newDBParameterGroupNameMessage smart constructor.

Constructors

DBParameterGroupNameMessage'

Fields

dbParameterGroupName :: Maybe Text

The name of the DB parameter group.

newDBParameterGroupNameMessage :: DBParameterGroupNameMessage

Create a value of DBParameterGroupNameMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbParameterGroupName:DBParameterGroupNameMessage', dbParameterGroupNameMessage_dbParameterGroupName - The name of the DB parameter group.

Response Lenses

dbParameterGroupNameMessage_dbParameterGroupName :: Lens' DBParameterGroupNameMessage (Maybe Text)

The name of the DB parameter group.