Amazonka.RDS.RestoreDBClusterFromSnapshot

Description

Creates a new DB cluster from a DB snapshot or DB cluster snapshot. This action only applies to Aurora DB clusters.

The target DB cluster is created from the source snapshot with a default configuration. If you don't specify a security group, the new DB cluster is associated with the default security group.

This action only restores the DB cluster, not the DB instances for that DB cluster. You must invoke the CreateDBInstance action to create DB instances for the restored DB cluster, specifying the identifier of the restored DB cluster in DBClusterIdentifier. You can create DB instances only after the RestoreDBClusterFromSnapshot action has completed and the DB cluster is available.

For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This action only applies to Aurora DB clusters.

Creating a Request

data RestoreDBClusterFromSnapshot

See: newRestoreDBClusterFromSnapshot smart constructor.

Constructors

RestoreDBClusterFromSnapshot'

Fields

engineVersion :: Maybe Text

The version of the database engine to use for the new DB cluster.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

If you aren't using the default engine version, then you must specify the engine version.

Aurora MySQL

Example: 5.6.10a, 5.6.mysql_aurora.1.19.2, 5.7.12, 5.7.mysql_aurora.2.04.5

Aurora PostgreSQL

Example: 9.6.3, 10.7

deletionProtection :: Maybe Bool

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

dbSubnetGroupName :: Maybe Text

The name of the DB subnet group to use for the new DB cluster.

Constraints: If supplied, must match the name of an existing DB subnet group.

Example: mySubnetgroup

domain :: Maybe Text

Specify the Active Directory directory ID to restore the DB cluster in. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

backtrackWindow :: Maybe Integer

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

availabilityZones :: Maybe [Text]

Provides the list of Availability Zones (AZs) where instances in the restored DB cluster can be created.

kmsKeyId :: Maybe Text

The Amazon Web Services KMS key identifier to use when restoring an encrypted DB cluster from a DB snapshot or DB cluster snapshot.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

When you don't specify a value for the KmsKeyId parameter, then the following occurs:

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier is encrypted, then the restored DB cluster is encrypted using the Amazon Web Services KMS CMK that was used to encrypt the DB snapshot or DB cluster snapshot.

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier isn't encrypted, then the restored DB cluster isn't encrypted.

vpcSecurityGroupIds :: Maybe [Text]

A list of VPC security groups that the new DB cluster will belong to.

databaseName :: Maybe Text

The database name for the restored DB cluster.

dbClusterParameterGroupName :: Maybe Text

The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, the default DB cluster parameter group for the specified engine is used.

Constraints:

If supplied, must match the name of an existing default DB cluster parameter group.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

engineMode :: Maybe Text

The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

For more information, see CreateDBCluster.

scalingConfiguration :: Maybe ScalingConfiguration

For DB clusters in serverless DB engine mode, the scaling properties of the DB cluster.

optionGroupName :: Maybe Text

The name of the option group to use for the restored DB cluster.

copyTagsToSnapshot :: Maybe Bool

A value that indicates whether to copy all tags from the restored DB cluster to snapshots of the restored DB cluster. The default is not to copy them.

domainIAMRoleName :: Maybe Text

Specify the name of the IAM role to be used when making API calls to the Directory Service.

tags :: Maybe [Tag]

The tags to be assigned to the restored DB cluster.

port :: Maybe Int

The port number on which the new DB cluster accepts connections.

Constraints: This value must be 1150-65535

Default: The same port as the original DB cluster.

enableIAMDatabaseAuthentication :: Maybe Bool

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

enableCloudwatchLogsExports :: Maybe [Text]

The list of logs that the restored DB cluster is to export to Amazon CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

dbClusterIdentifier :: Text

The name of the DB cluster to create from the DB snapshot or DB cluster snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

snapshotIdentifier :: Text

The identifier for the DB snapshot or DB cluster snapshot to restore from.

You can use either the name or the Amazon Resource Name (ARN) to specify a DB cluster snapshot. However, you can use only the ARN to specify a DB snapshot.

Constraints:

Must match the identifier of an existing Snapshot.

engine :: Text

The database engine to use for the new DB cluster.

Default: The same as source

Constraint: Must be compatible with the engine of the source

newRestoreDBClusterFromSnapshot

Arguments

:: Text
$sel:dbClusterIdentifier:RestoreDBClusterFromSnapshot'

-> Text
$sel:snapshotIdentifier:RestoreDBClusterFromSnapshot'

-> Text
$sel:engine:RestoreDBClusterFromSnapshot'

-> RestoreDBClusterFromSnapshot

Create a value of RestoreDBClusterFromSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_engineVersion - The version of the database engine to use for the new DB cluster.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

If you aren't using the default engine version, then you must specify the engine version.

Aurora MySQL

Example: 5.6.10a, 5.6.mysql_aurora.1.19.2, 5.7.12, 5.7.mysql_aurora.2.04.5

Aurora PostgreSQL

Example: 9.6.3, 10.7

$sel:deletionProtection:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_deletionProtection - A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

$sel:dbSubnetGroupName:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_dbSubnetGroupName - The name of the DB subnet group to use for the new DB cluster.

Constraints: If supplied, must match the name of an existing DB subnet group.

Example: mySubnetgroup

$sel:domain:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_domain - Specify the Active Directory directory ID to restore the DB cluster in. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

$sel:backtrackWindow:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_backtrackWindow - The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

$sel:availabilityZones:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_availabilityZones - Provides the list of Availability Zones (AZs) where instances in the restored DB cluster can be created.

$sel:kmsKeyId:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_kmsKeyId - The Amazon Web Services KMS key identifier to use when restoring an encrypted DB cluster from a DB snapshot or DB cluster snapshot.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

When you don't specify a value for the KmsKeyId parameter, then the following occurs:

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier is encrypted, then the restored DB cluster is encrypted using the Amazon Web Services KMS CMK that was used to encrypt the DB snapshot or DB cluster snapshot.

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier isn't encrypted, then the restored DB cluster isn't encrypted.

$sel:vpcSecurityGroupIds:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_vpcSecurityGroupIds - A list of VPC security groups that the new DB cluster will belong to.

$sel:databaseName:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_databaseName - The database name for the restored DB cluster.

$sel:dbClusterParameterGroupName:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_dbClusterParameterGroupName - The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, the default DB cluster parameter group for the specified engine is used.

Constraints:

If supplied, must match the name of an existing default DB cluster parameter group.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

$sel:engineMode:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_engineMode - The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

For more information, see CreateDBCluster.

$sel:scalingConfiguration:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_scalingConfiguration - For DB clusters in serverless DB engine mode, the scaling properties of the DB cluster.

$sel:optionGroupName:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_optionGroupName - The name of the option group to use for the restored DB cluster.

$sel:copyTagsToSnapshot:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_copyTagsToSnapshot - A value that indicates whether to copy all tags from the restored DB cluster to snapshots of the restored DB cluster. The default is not to copy them.

$sel:domainIAMRoleName:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_domainIAMRoleName - Specify the name of the IAM role to be used when making API calls to the Directory Service.

$sel:tags:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_tags - The tags to be assigned to the restored DB cluster.

$sel:port:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_port - The port number on which the new DB cluster accepts connections.

Constraints: This value must be 1150-65535

Default: The same port as the original DB cluster.

$sel:enableIAMDatabaseAuthentication:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_enableIAMDatabaseAuthentication - A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

$sel:enableCloudwatchLogsExports:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_enableCloudwatchLogsExports - The list of logs that the restored DB cluster is to export to Amazon CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

$sel:dbClusterIdentifier:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_dbClusterIdentifier - The name of the DB cluster to create from the DB snapshot or DB cluster snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

$sel:snapshotIdentifier:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_snapshotIdentifier - The identifier for the DB snapshot or DB cluster snapshot to restore from.

You can use either the name or the Amazon Resource Name (ARN) to specify a DB cluster snapshot. However, you can use only the ARN to specify a DB snapshot.

Constraints:

Must match the identifier of an existing Snapshot.

$sel:engine:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_engine - The database engine to use for the new DB cluster.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Request Lenses

restoreDBClusterFromSnapshot_engineVersion :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The version of the database engine to use for the new DB cluster.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

If you aren't using the default engine version, then you must specify the engine version.

Aurora MySQL

Example: 5.6.10a, 5.6.mysql_aurora.1.19.2, 5.7.12, 5.7.mysql_aurora.2.04.5

Aurora PostgreSQL

Example: 9.6.3, 10.7

restoreDBClusterFromSnapshot_deletionProtection :: Lens' RestoreDBClusterFromSnapshot (Maybe Bool)

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

restoreDBClusterFromSnapshot_dbSubnetGroupName :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The name of the DB subnet group to use for the new DB cluster.

Constraints: If supplied, must match the name of an existing DB subnet group.

Example: mySubnetgroup

restoreDBClusterFromSnapshot_domain :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

Specify the Active Directory directory ID to restore the DB cluster in. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

restoreDBClusterFromSnapshot_backtrackWindow :: Lens' RestoreDBClusterFromSnapshot (Maybe Integer)

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

restoreDBClusterFromSnapshot_availabilityZones :: Lens' RestoreDBClusterFromSnapshot (Maybe [Text])

Provides the list of Availability Zones (AZs) where instances in the restored DB cluster can be created.

restoreDBClusterFromSnapshot_kmsKeyId :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The Amazon Web Services KMS key identifier to use when restoring an encrypted DB cluster from a DB snapshot or DB cluster snapshot.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

When you don't specify a value for the KmsKeyId parameter, then the following occurs:

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier is encrypted, then the restored DB cluster is encrypted using the Amazon Web Services KMS CMK that was used to encrypt the DB snapshot or DB cluster snapshot.

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier isn't encrypted, then the restored DB cluster isn't encrypted.

restoreDBClusterFromSnapshot_vpcSecurityGroupIds :: Lens' RestoreDBClusterFromSnapshot (Maybe [Text])

A list of VPC security groups that the new DB cluster will belong to.

restoreDBClusterFromSnapshot_databaseName :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The database name for the restored DB cluster.

restoreDBClusterFromSnapshot_dbClusterParameterGroupName :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, the default DB cluster parameter group for the specified engine is used.

Constraints:

If supplied, must match the name of an existing default DB cluster parameter group.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

restoreDBClusterFromSnapshot_engineMode :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

For more information, see CreateDBCluster.

restoreDBClusterFromSnapshot_scalingConfiguration :: Lens' RestoreDBClusterFromSnapshot (Maybe ScalingConfiguration)

For DB clusters in serverless DB engine mode, the scaling properties of the DB cluster.

restoreDBClusterFromSnapshot_optionGroupName :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The name of the option group to use for the restored DB cluster.

restoreDBClusterFromSnapshot_copyTagsToSnapshot :: Lens' RestoreDBClusterFromSnapshot (Maybe Bool)

A value that indicates whether to copy all tags from the restored DB cluster to snapshots of the restored DB cluster. The default is not to copy them.

restoreDBClusterFromSnapshot_domainIAMRoleName :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

Specify the name of the IAM role to be used when making API calls to the Directory Service.

restoreDBClusterFromSnapshot_tags :: Lens' RestoreDBClusterFromSnapshot (Maybe [Tag])

The tags to be assigned to the restored DB cluster.

restoreDBClusterFromSnapshot_port :: Lens' RestoreDBClusterFromSnapshot (Maybe Int)

The port number on which the new DB cluster accepts connections.

Constraints: This value must be 1150-65535

Default: The same port as the original DB cluster.

restoreDBClusterFromSnapshot_enableIAMDatabaseAuthentication :: Lens' RestoreDBClusterFromSnapshot (Maybe Bool)

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

restoreDBClusterFromSnapshot_enableCloudwatchLogsExports :: Lens' RestoreDBClusterFromSnapshot (Maybe [Text])

The list of logs that the restored DB cluster is to export to Amazon CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

restoreDBClusterFromSnapshot_dbClusterIdentifier :: Lens' RestoreDBClusterFromSnapshot Text

The name of the DB cluster to create from the DB snapshot or DB cluster snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

restoreDBClusterFromSnapshot_snapshotIdentifier :: Lens' RestoreDBClusterFromSnapshot Text

The identifier for the DB snapshot or DB cluster snapshot to restore from.

You can use either the name or the Amazon Resource Name (ARN) to specify a DB cluster snapshot. However, you can use only the ARN to specify a DB snapshot.

Constraints:

Must match the identifier of an existing Snapshot.

restoreDBClusterFromSnapshot_engine :: Lens' RestoreDBClusterFromSnapshot Text

The database engine to use for the new DB cluster.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Destructuring the Response

data RestoreDBClusterFromSnapshotResponse

See: newRestoreDBClusterFromSnapshotResponse smart constructor.

Constructors

RestoreDBClusterFromSnapshotResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newRestoreDBClusterFromSnapshotResponse

Arguments

:: Int
$sel:httpStatus:RestoreDBClusterFromSnapshotResponse'

-> RestoreDBClusterFromSnapshotResponse

Create a value of RestoreDBClusterFromSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:RestoreDBClusterFromSnapshotResponse', restoreDBClusterFromSnapshotResponse_dbCluster - Undocumented member.

$sel:httpStatus:RestoreDBClusterFromSnapshotResponse', restoreDBClusterFromSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

restoreDBClusterFromSnapshotResponse_dbCluster :: Lens' RestoreDBClusterFromSnapshotResponse (Maybe DBCluster)

Undocumented member.

restoreDBClusterFromSnapshotResponse_httpStatus :: Lens' RestoreDBClusterFromSnapshotResponse Int

The response's http status code.