Amazonka.RDS.RestoreDBClusterToPointInTime

Description

Restores a DB cluster to an arbitrary point in time. Users can restore to any point in time before LatestRestorableTime for up to BackupRetentionPeriod days. The target DB cluster is created from the source DB cluster with the same configuration as the original DB cluster, except that the new DB cluster is created with the default DB security group.

This action only restores the DB cluster, not the DB instances for that DB cluster. You must invoke the CreateDBInstance action to create DB instances for the restored DB cluster, specifying the identifier of the restored DB cluster in DBClusterIdentifier. You can create DB instances only after the RestoreDBClusterToPointInTime action has completed and the DB cluster is available.

For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This action only applies to Aurora DB clusters.

Creating a Request

data RestoreDBClusterToPointInTime

See: newRestoreDBClusterToPointInTime smart constructor.

Constructors

RestoreDBClusterToPointInTime'

Fields

deletionProtection :: Maybe Bool

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

useLatestRestorableTime :: Maybe Bool

A value that indicates whether to restore the DB cluster to the latest restorable backup time. By default, the DB cluster isn't restored to the latest restorable backup time.

Constraints: Can't be specified if RestoreToTime parameter is provided.

dbSubnetGroupName :: Maybe Text

The DB subnet group name to use for the new DB cluster.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

domain :: Maybe Text

Specify the Active Directory directory ID to restore the DB cluster in. The domain must be created prior to this operation.

For Amazon Aurora DB clusters, Amazon RDS can use Kerberos Authentication to authenticate users that connect to the DB cluster. For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

backtrackWindow :: Maybe Integer

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

kmsKeyId :: Maybe Text

The Amazon Web Services KMS key identifier to use when restoring an encrypted DB cluster from an encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

You can restore to a new DB cluster and encrypt the new DB cluster with a Amazon Web Services KMS CMK that is different than the Amazon Web Services KMS key used to encrypt the source DB cluster. The new DB cluster is encrypted with the Amazon Web Services KMS CMK identified by the KmsKeyId parameter.

If you don't specify a value for the KmsKeyId parameter, then the following occurs:

If the DB cluster is encrypted, then the restored DB cluster is encrypted using the Amazon Web Services KMS CMK that was used to encrypt the source DB cluster.

If the DB cluster isn't encrypted, then the restored DB cluster isn't encrypted.

If DBClusterIdentifier refers to a DB cluster that isn't encrypted, then the restore request is rejected.

vpcSecurityGroupIds :: Maybe [Text]

A list of VPC security groups that the new DB cluster belongs to.

dbClusterParameterGroupName :: Maybe Text

The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, the default DB cluster parameter group for the specified engine is used.

Constraints:

If supplied, must match the name of an existing DB cluster parameter group.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

engineMode :: Maybe Text

The engine mode of the new cluster. Specify provisioned or serverless, depending on the type of the cluster you are creating. You can create an Aurora Serverless clone from a provisioned cluster, or a provisioned clone from an Aurora Serverless cluster. To create a clone that is an Aurora Serverless cluster, the original cluster must be an Aurora Serverless cluster or an encrypted provisioned cluster.

scalingConfiguration :: Maybe ScalingConfiguration

For DB clusters in serverless DB engine mode, the scaling properties of the DB cluster.

restoreType :: Maybe Text

The type of restore to be performed. You can specify one of the following values:

full-copy - The new DB cluster is restored as a full copy of the source DB cluster.

copy-on-write - The new DB cluster is restored as a clone of the source DB cluster.

Constraints: You can't specify copy-on-write if the engine version of the source DB cluster is earlier than 1.11.

If you don't specify a RestoreType value, then the new DB cluster is restored as a full copy of the source DB cluster.

optionGroupName :: Maybe Text

The name of the option group for the new DB cluster.

copyTagsToSnapshot :: Maybe Bool

A value that indicates whether to copy all tags from the restored DB cluster to snapshots of the restored DB cluster. The default is not to copy them.

restoreToTime :: Maybe ISO8601

The date and time to restore the DB cluster to.

Valid Values: Value must be a time in Universal Coordinated Time (UTC) format

Constraints:

Must be before the latest restorable time for the DB instance

Must be specified if UseLatestRestorableTime parameter isn't provided

Can't be specified if the UseLatestRestorableTime parameter is enabled

Can't be specified if the RestoreType parameter is copy-on-write

Example: 2015-03-07T23:45:00Z

domainIAMRoleName :: Maybe Text

Specify the name of the IAM role to be used when making API calls to the Directory Service.

tags :: Maybe [Tag]

port :: Maybe Int

The port number on which the new DB cluster accepts connections.

Constraints: A value from 1150-65535.

Default: The default port for the engine.

enableIAMDatabaseAuthentication :: Maybe Bool

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

enableCloudwatchLogsExports :: Maybe [Text]

The list of logs that the restored DB cluster is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

dbClusterIdentifier :: Text

The name of the new DB cluster to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

sourceDBClusterIdentifier :: Text

The identifier of the source DB cluster from which to restore.

Constraints:

Must match the identifier of an existing DBCluster.

newRestoreDBClusterToPointInTime

Arguments

:: Text
$sel:dbClusterIdentifier:RestoreDBClusterToPointInTime'

-> Text
$sel:sourceDBClusterIdentifier:RestoreDBClusterToPointInTime'

-> RestoreDBClusterToPointInTime

Create a value of RestoreDBClusterToPointInTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletionProtection:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_deletionProtection - A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

$sel:useLatestRestorableTime:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_useLatestRestorableTime - A value that indicates whether to restore the DB cluster to the latest restorable backup time. By default, the DB cluster isn't restored to the latest restorable backup time.

Constraints: Can't be specified if RestoreToTime parameter is provided.

$sel:dbSubnetGroupName:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_dbSubnetGroupName - The DB subnet group name to use for the new DB cluster.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

$sel:domain:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_domain - Specify the Active Directory directory ID to restore the DB cluster in. The domain must be created prior to this operation.

For Amazon Aurora DB clusters, Amazon RDS can use Kerberos Authentication to authenticate users that connect to the DB cluster. For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

$sel:backtrackWindow:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_backtrackWindow - The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

$sel:kmsKeyId:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_kmsKeyId - The Amazon Web Services KMS key identifier to use when restoring an encrypted DB cluster from an encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

You can restore to a new DB cluster and encrypt the new DB cluster with a Amazon Web Services KMS CMK that is different than the Amazon Web Services KMS key used to encrypt the source DB cluster. The new DB cluster is encrypted with the Amazon Web Services KMS CMK identified by the KmsKeyId parameter.

If you don't specify a value for the KmsKeyId parameter, then the following occurs:

If the DB cluster is encrypted, then the restored DB cluster is encrypted using the Amazon Web Services KMS CMK that was used to encrypt the source DB cluster.

If the DB cluster isn't encrypted, then the restored DB cluster isn't encrypted.

If DBClusterIdentifier refers to a DB cluster that isn't encrypted, then the restore request is rejected.

$sel:vpcSecurityGroupIds:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_vpcSecurityGroupIds - A list of VPC security groups that the new DB cluster belongs to.

$sel:dbClusterParameterGroupName:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_dbClusterParameterGroupName - The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, the default DB cluster parameter group for the specified engine is used.

Constraints:

If supplied, must match the name of an existing DB cluster parameter group.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

$sel:engineMode:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_engineMode - The engine mode of the new cluster. Specify provisioned or serverless, depending on the type of the cluster you are creating. You can create an Aurora Serverless clone from a provisioned cluster, or a provisioned clone from an Aurora Serverless cluster. To create a clone that is an Aurora Serverless cluster, the original cluster must be an Aurora Serverless cluster or an encrypted provisioned cluster.

$sel:scalingConfiguration:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_scalingConfiguration - For DB clusters in serverless DB engine mode, the scaling properties of the DB cluster.

$sel:restoreType:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_restoreType - The type of restore to be performed. You can specify one of the following values:

full-copy - The new DB cluster is restored as a full copy of the source DB cluster.

copy-on-write - The new DB cluster is restored as a clone of the source DB cluster.

Constraints: You can't specify copy-on-write if the engine version of the source DB cluster is earlier than 1.11.

If you don't specify a RestoreType value, then the new DB cluster is restored as a full copy of the source DB cluster.

$sel:optionGroupName:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_optionGroupName - The name of the option group for the new DB cluster.

$sel:copyTagsToSnapshot:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_copyTagsToSnapshot - A value that indicates whether to copy all tags from the restored DB cluster to snapshots of the restored DB cluster. The default is not to copy them.

$sel:restoreToTime:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_restoreToTime - The date and time to restore the DB cluster to.

Valid Values: Value must be a time in Universal Coordinated Time (UTC) format

Constraints:

Must be before the latest restorable time for the DB instance

Must be specified if UseLatestRestorableTime parameter isn't provided

Can't be specified if the UseLatestRestorableTime parameter is enabled

Can't be specified if the RestoreType parameter is copy-on-write

Example: 2015-03-07T23:45:00Z

$sel:domainIAMRoleName:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_domainIAMRoleName - Specify the name of the IAM role to be used when making API calls to the Directory Service.

$sel:tags:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_tags - Undocumented member.

$sel:port:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_port - The port number on which the new DB cluster accepts connections.

Constraints: A value from 1150-65535.

Default: The default port for the engine.

$sel:enableIAMDatabaseAuthentication:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_enableIAMDatabaseAuthentication - A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

$sel:enableCloudwatchLogsExports:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_enableCloudwatchLogsExports - The list of logs that the restored DB cluster is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

$sel:dbClusterIdentifier:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_dbClusterIdentifier - The name of the new DB cluster to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

$sel:sourceDBClusterIdentifier:RestoreDBClusterToPointInTime', restoreDBClusterToPointInTime_sourceDBClusterIdentifier - The identifier of the source DB cluster from which to restore.

Constraints:

Must match the identifier of an existing DBCluster.

Request Lenses

restoreDBClusterToPointInTime_deletionProtection :: Lens' RestoreDBClusterToPointInTime (Maybe Bool)

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

restoreDBClusterToPointInTime_useLatestRestorableTime :: Lens' RestoreDBClusterToPointInTime (Maybe Bool)

A value that indicates whether to restore the DB cluster to the latest restorable backup time. By default, the DB cluster isn't restored to the latest restorable backup time.

Constraints: Can't be specified if RestoreToTime parameter is provided.

restoreDBClusterToPointInTime_dbSubnetGroupName :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

The DB subnet group name to use for the new DB cluster.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

restoreDBClusterToPointInTime_domain :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

Specify the Active Directory directory ID to restore the DB cluster in. The domain must be created prior to this operation.

For Amazon Aurora DB clusters, Amazon RDS can use Kerberos Authentication to authenticate users that connect to the DB cluster. For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

restoreDBClusterToPointInTime_backtrackWindow :: Lens' RestoreDBClusterToPointInTime (Maybe Integer)

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

restoreDBClusterToPointInTime_kmsKeyId :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

The Amazon Web Services KMS key identifier to use when restoring an encrypted DB cluster from an encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

You can restore to a new DB cluster and encrypt the new DB cluster with a Amazon Web Services KMS CMK that is different than the Amazon Web Services KMS key used to encrypt the source DB cluster. The new DB cluster is encrypted with the Amazon Web Services KMS CMK identified by the KmsKeyId parameter.

If you don't specify a value for the KmsKeyId parameter, then the following occurs:

If the DB cluster is encrypted, then the restored DB cluster is encrypted using the Amazon Web Services KMS CMK that was used to encrypt the source DB cluster.

If the DB cluster isn't encrypted, then the restored DB cluster isn't encrypted.

If DBClusterIdentifier refers to a DB cluster that isn't encrypted, then the restore request is rejected.

restoreDBClusterToPointInTime_vpcSecurityGroupIds :: Lens' RestoreDBClusterToPointInTime (Maybe [Text])

A list of VPC security groups that the new DB cluster belongs to.

restoreDBClusterToPointInTime_dbClusterParameterGroupName :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, the default DB cluster parameter group for the specified engine is used.

Constraints:

If supplied, must match the name of an existing DB cluster parameter group.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

restoreDBClusterToPointInTime_engineMode :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

The engine mode of the new cluster. Specify provisioned or serverless, depending on the type of the cluster you are creating. You can create an Aurora Serverless clone from a provisioned cluster, or a provisioned clone from an Aurora Serverless cluster. To create a clone that is an Aurora Serverless cluster, the original cluster must be an Aurora Serverless cluster or an encrypted provisioned cluster.

restoreDBClusterToPointInTime_scalingConfiguration :: Lens' RestoreDBClusterToPointInTime (Maybe ScalingConfiguration)

For DB clusters in serverless DB engine mode, the scaling properties of the DB cluster.

restoreDBClusterToPointInTime_restoreType :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

The type of restore to be performed. You can specify one of the following values:

full-copy - The new DB cluster is restored as a full copy of the source DB cluster.

copy-on-write - The new DB cluster is restored as a clone of the source DB cluster.

Constraints: You can't specify copy-on-write if the engine version of the source DB cluster is earlier than 1.11.

If you don't specify a RestoreType value, then the new DB cluster is restored as a full copy of the source DB cluster.

restoreDBClusterToPointInTime_optionGroupName :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

The name of the option group for the new DB cluster.

restoreDBClusterToPointInTime_copyTagsToSnapshot :: Lens' RestoreDBClusterToPointInTime (Maybe Bool)

A value that indicates whether to copy all tags from the restored DB cluster to snapshots of the restored DB cluster. The default is not to copy them.

restoreDBClusterToPointInTime_restoreToTime :: Lens' RestoreDBClusterToPointInTime (Maybe UTCTime)

The date and time to restore the DB cluster to.

Valid Values: Value must be a time in Universal Coordinated Time (UTC) format

Constraints:

Must be before the latest restorable time for the DB instance

Must be specified if UseLatestRestorableTime parameter isn't provided

Can't be specified if the UseLatestRestorableTime parameter is enabled

Can't be specified if the RestoreType parameter is copy-on-write

Example: 2015-03-07T23:45:00Z

restoreDBClusterToPointInTime_domainIAMRoleName :: Lens' RestoreDBClusterToPointInTime (Maybe Text)

Specify the name of the IAM role to be used when making API calls to the Directory Service.

restoreDBClusterToPointInTime_tags :: Lens' RestoreDBClusterToPointInTime (Maybe [Tag])

Undocumented member.

restoreDBClusterToPointInTime_port :: Lens' RestoreDBClusterToPointInTime (Maybe Int)

The port number on which the new DB cluster accepts connections.

Constraints: A value from 1150-65535.

Default: The default port for the engine.

restoreDBClusterToPointInTime_enableIAMDatabaseAuthentication :: Lens' RestoreDBClusterToPointInTime (Maybe Bool)

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

restoreDBClusterToPointInTime_enableCloudwatchLogsExports :: Lens' RestoreDBClusterToPointInTime (Maybe [Text])

The list of logs that the restored DB cluster is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

restoreDBClusterToPointInTime_dbClusterIdentifier :: Lens' RestoreDBClusterToPointInTime Text

The name of the new DB cluster to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

restoreDBClusterToPointInTime_sourceDBClusterIdentifier :: Lens' RestoreDBClusterToPointInTime Text

The identifier of the source DB cluster from which to restore.

Constraints:

Must match the identifier of an existing DBCluster.

Destructuring the Response

data RestoreDBClusterToPointInTimeResponse

See: newRestoreDBClusterToPointInTimeResponse smart constructor.

Constructors

RestoreDBClusterToPointInTimeResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newRestoreDBClusterToPointInTimeResponse

Arguments

:: Int
$sel:httpStatus:RestoreDBClusterToPointInTimeResponse'

-> RestoreDBClusterToPointInTimeResponse

Create a value of RestoreDBClusterToPointInTimeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:RestoreDBClusterToPointInTimeResponse', restoreDBClusterToPointInTimeResponse_dbCluster - Undocumented member.

$sel:httpStatus:RestoreDBClusterToPointInTimeResponse', restoreDBClusterToPointInTimeResponse_httpStatus - The response's http status code.

Response Lenses

restoreDBClusterToPointInTimeResponse_dbCluster :: Lens' RestoreDBClusterToPointInTimeResponse (Maybe DBCluster)

Undocumented member.

restoreDBClusterToPointInTimeResponse_httpStatus :: Lens' RestoreDBClusterToPointInTimeResponse Int

The response's http status code.