Amazonka.RDS.StartActivityStream

Description

Starts a database activity stream to monitor activity on the database. For more information, see Database Activity Streams in the Amazon Aurora User Guide.

Creating a Request

data StartActivityStream

See: newStartActivityStream smart constructor.

Constructors

StartActivityStream'

Fields

engineNativeAuditFieldsIncluded :: Maybe Bool

Specifies whether the database activity stream includes engine-native audit fields. This option only applies to an Oracle DB instance. By default, no engine-native audit fields are included.

applyImmediately :: Maybe Bool

Specifies whether or not the database activity stream is to start as soon as possible, regardless of the maintenance window for the database.

resourceArn :: Text

The Amazon Resource Name (ARN) of the DB cluster, for example, arn:aws:rds:us-east-1:12345667890:cluster:das-cluster.

mode :: ActivityStreamMode

Specifies the mode of the database activity stream. Database events such as a change or access generate an activity stream event. The database session can handle these events either synchronously or asynchronously.

kmsKeyId :: Text

The Amazon Web Services KMS key identifier for encrypting messages in the database activity stream. The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

newStartActivityStream

Arguments

:: Text
$sel:resourceArn:StartActivityStream'

-> ActivityStreamMode
$sel:mode:StartActivityStream'

-> Text
$sel:kmsKeyId:StartActivityStream'

-> StartActivityStream

Create a value of StartActivityStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineNativeAuditFieldsIncluded:StartActivityStream', startActivityStream_engineNativeAuditFieldsIncluded - Specifies whether the database activity stream includes engine-native audit fields. This option only applies to an Oracle DB instance. By default, no engine-native audit fields are included.

$sel:applyImmediately:StartActivityStream', startActivityStream_applyImmediately - Specifies whether or not the database activity stream is to start as soon as possible, regardless of the maintenance window for the database.

$sel:resourceArn:StartActivityStream', startActivityStream_resourceArn - The Amazon Resource Name (ARN) of the DB cluster, for example, arn:aws:rds:us-east-1:12345667890:cluster:das-cluster.

$sel:mode:StartActivityStream', startActivityStream_mode - Specifies the mode of the database activity stream. Database events such as a change or access generate an activity stream event. The database session can handle these events either synchronously or asynchronously.

$sel:kmsKeyId:StartActivityStream', startActivityStream_kmsKeyId - The Amazon Web Services KMS key identifier for encrypting messages in the database activity stream. The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

Request Lenses

startActivityStream_engineNativeAuditFieldsIncluded :: Lens' StartActivityStream (Maybe Bool)

Specifies whether the database activity stream includes engine-native audit fields. This option only applies to an Oracle DB instance. By default, no engine-native audit fields are included.

startActivityStream_applyImmediately :: Lens' StartActivityStream (Maybe Bool)

Specifies whether or not the database activity stream is to start as soon as possible, regardless of the maintenance window for the database.

startActivityStream_resourceArn :: Lens' StartActivityStream Text

The Amazon Resource Name (ARN) of the DB cluster, for example, arn:aws:rds:us-east-1:12345667890:cluster:das-cluster.

startActivityStream_mode :: Lens' StartActivityStream ActivityStreamMode

Specifies the mode of the database activity stream. Database events such as a change or access generate an activity stream event. The database session can handle these events either synchronously or asynchronously.

startActivityStream_kmsKeyId :: Lens' StartActivityStream Text

The Amazon Web Services KMS key identifier for encrypting messages in the database activity stream. The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

Destructuring the Response

data StartActivityStreamResponse

See: newStartActivityStreamResponse smart constructor.

Constructors

StartActivityStreamResponse'

Fields

status :: Maybe ActivityStreamStatus

The status of the database activity stream.

kinesisStreamName :: Maybe Text

The name of the Amazon Kinesis data stream to be used for the database activity stream.

engineNativeAuditFieldsIncluded :: Maybe Bool

Indicates whether engine-native audit fields are included in the database activity stream.

mode :: Maybe ActivityStreamMode

The mode of the database activity stream.

kmsKeyId :: Maybe Text

The Amazon Web Services KMS key identifier for encryption of messages in the database activity stream.

applyImmediately :: Maybe Bool

Indicates whether or not the database activity stream will start as soon as possible, regardless of the maintenance window for the database.

httpStatus :: Int

The response's http status code.

newStartActivityStreamResponse

Arguments

:: Int
$sel:httpStatus:StartActivityStreamResponse'

-> StartActivityStreamResponse

Create a value of StartActivityStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StartActivityStreamResponse', startActivityStreamResponse_status - The status of the database activity stream.

$sel:kinesisStreamName:StartActivityStreamResponse', startActivityStreamResponse_kinesisStreamName - The name of the Amazon Kinesis data stream to be used for the database activity stream.

$sel:engineNativeAuditFieldsIncluded:StartActivityStream', startActivityStreamResponse_engineNativeAuditFieldsIncluded - Indicates whether engine-native audit fields are included in the database activity stream.

$sel:mode:StartActivityStream', startActivityStreamResponse_mode - The mode of the database activity stream.

$sel:kmsKeyId:StartActivityStream', startActivityStreamResponse_kmsKeyId - The Amazon Web Services KMS key identifier for encryption of messages in the database activity stream.

$sel:applyImmediately:StartActivityStream', startActivityStreamResponse_applyImmediately - Indicates whether or not the database activity stream will start as soon as possible, regardless of the maintenance window for the database.

$sel:httpStatus:StartActivityStreamResponse', startActivityStreamResponse_httpStatus - The response's http status code.

Response Lenses

startActivityStreamResponse_status :: Lens' StartActivityStreamResponse (Maybe ActivityStreamStatus)

The status of the database activity stream.

startActivityStreamResponse_kinesisStreamName :: Lens' StartActivityStreamResponse (Maybe Text)

The name of the Amazon Kinesis data stream to be used for the database activity stream.

startActivityStreamResponse_engineNativeAuditFieldsIncluded :: Lens' StartActivityStreamResponse (Maybe Bool)

Indicates whether engine-native audit fields are included in the database activity stream.

startActivityStreamResponse_mode :: Lens' StartActivityStreamResponse (Maybe ActivityStreamMode)

The mode of the database activity stream.

startActivityStreamResponse_kmsKeyId :: Lens' StartActivityStreamResponse (Maybe Text)

The Amazon Web Services KMS key identifier for encryption of messages in the database activity stream.

startActivityStreamResponse_applyImmediately :: Lens' StartActivityStreamResponse (Maybe Bool)

Indicates whether or not the database activity stream will start as soon as possible, regardless of the maintenance window for the database.

startActivityStreamResponse_httpStatus :: Lens' StartActivityStreamResponse Int

The response's http status code.