Amazonka.RDS.StartExportTask

Description

Starts an export of a snapshot to Amazon S3. The provided IAM role must have access to the S3 bucket.

Creating a Request

data StartExportTask

See: newStartExportTask smart constructor.

Constructors

StartExportTask'

Fields

exportOnly :: Maybe [Text]

The data to be exported from the snapshot. If this parameter is not provided, all the snapshot data is exported. Valid values are the following:

database - Export all the data from a specified database.

database.table table-name - Export a table of the snapshot. This format is valid only for RDS for MySQL, RDS for MariaDB, and Aurora MySQL.

database.schema schema-name - Export a database schema of the snapshot. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

database.schema.table table-name - Export a table of the database schema. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

s3Prefix :: Maybe Text

The Amazon S3 bucket prefix to use as the file name and path of the exported snapshot.

exportTaskIdentifier :: Text

A unique identifier for the snapshot export task. This ID isn't an identifier for the Amazon S3 bucket where the snapshot is to be exported to.

sourceArn :: Text

The Amazon Resource Name (ARN) of the snapshot to export to Amazon S3.

s3BucketName :: Text

The name of the Amazon S3 bucket to export the snapshot to.

iamRoleArn :: Text

The name of the IAM role to use for writing to the Amazon S3 bucket when exporting a snapshot.

kmsKeyId :: Text

The ID of the Amazon Web Services KMS customer master key (CMK) to use to encrypt the snapshot exported to Amazon S3. The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). The caller of this operation must be authorized to execute the following operations. These can be set in the Amazon Web Services KMS key policy:

GrantOperation.Encrypt

GrantOperation.Decrypt

GrantOperation.GenerateDataKey

GrantOperation.GenerateDataKeyWithoutPlaintext

GrantOperation.ReEncryptFrom

GrantOperation.ReEncryptTo

GrantOperation.CreateGrant

GrantOperation.DescribeKey

GrantOperation.RetireGrant

newStartExportTask

Arguments

:: Text
$sel:exportTaskIdentifier:StartExportTask'

-> Text
$sel:sourceArn:StartExportTask'

-> Text
$sel:s3BucketName:StartExportTask'

-> Text
$sel:iamRoleArn:StartExportTask'

-> Text
$sel:kmsKeyId:StartExportTask'

-> StartExportTask

Create a value of StartExportTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exportOnly:StartExportTask', startExportTask_exportOnly - The data to be exported from the snapshot. If this parameter is not provided, all the snapshot data is exported. Valid values are the following:

database - Export all the data from a specified database.

database.table table-name - Export a table of the snapshot. This format is valid only for RDS for MySQL, RDS for MariaDB, and Aurora MySQL.

database.schema schema-name - Export a database schema of the snapshot. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

database.schema.table table-name - Export a table of the database schema. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

$sel:s3Prefix:StartExportTask', startExportTask_s3Prefix - The Amazon S3 bucket prefix to use as the file name and path of the exported snapshot.

$sel:exportTaskIdentifier:StartExportTask', startExportTask_exportTaskIdentifier - A unique identifier for the snapshot export task. This ID isn't an identifier for the Amazon S3 bucket where the snapshot is to be exported to.

$sel:sourceArn:StartExportTask', startExportTask_sourceArn - The Amazon Resource Name (ARN) of the snapshot to export to Amazon S3.

$sel:s3BucketName:StartExportTask', startExportTask_s3BucketName - The name of the Amazon S3 bucket to export the snapshot to.

$sel:iamRoleArn:StartExportTask', startExportTask_iamRoleArn - The name of the IAM role to use for writing to the Amazon S3 bucket when exporting a snapshot.

$sel:kmsKeyId:StartExportTask', startExportTask_kmsKeyId - The ID of the Amazon Web Services KMS customer master key (CMK) to use to encrypt the snapshot exported to Amazon S3. The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). The caller of this operation must be authorized to execute the following operations. These can be set in the Amazon Web Services KMS key policy:

GrantOperation.Encrypt

GrantOperation.Decrypt

GrantOperation.GenerateDataKey

GrantOperation.GenerateDataKeyWithoutPlaintext

GrantOperation.ReEncryptFrom

GrantOperation.ReEncryptTo

GrantOperation.CreateGrant

GrantOperation.DescribeKey

GrantOperation.RetireGrant

Request Lenses

startExportTask_exportOnly :: Lens' StartExportTask (Maybe [Text])

The data to be exported from the snapshot. If this parameter is not provided, all the snapshot data is exported. Valid values are the following:

database - Export all the data from a specified database.

database.table table-name - Export a table of the snapshot. This format is valid only for RDS for MySQL, RDS for MariaDB, and Aurora MySQL.

database.schema schema-name - Export a database schema of the snapshot. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

database.schema.table table-name - Export a table of the database schema. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

startExportTask_s3Prefix :: Lens' StartExportTask (Maybe Text)

The Amazon S3 bucket prefix to use as the file name and path of the exported snapshot.

startExportTask_exportTaskIdentifier :: Lens' StartExportTask Text

A unique identifier for the snapshot export task. This ID isn't an identifier for the Amazon S3 bucket where the snapshot is to be exported to.

startExportTask_sourceArn :: Lens' StartExportTask Text

The Amazon Resource Name (ARN) of the snapshot to export to Amazon S3.

startExportTask_s3BucketName :: Lens' StartExportTask Text

The name of the Amazon S3 bucket to export the snapshot to.

startExportTask_iamRoleArn :: Lens' StartExportTask Text

The name of the IAM role to use for writing to the Amazon S3 bucket when exporting a snapshot.

startExportTask_kmsKeyId :: Lens' StartExportTask Text

The ID of the Amazon Web Services KMS customer master key (CMK) to use to encrypt the snapshot exported to Amazon S3. The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). The caller of this operation must be authorized to execute the following operations. These can be set in the Amazon Web Services KMS key policy:

GrantOperation.Encrypt

GrantOperation.Decrypt

GrantOperation.GenerateDataKey

GrantOperation.GenerateDataKeyWithoutPlaintext

GrantOperation.ReEncryptFrom

GrantOperation.ReEncryptTo

GrantOperation.CreateGrant

GrantOperation.DescribeKey

GrantOperation.RetireGrant

Destructuring the Response

data ExportTask

Contains the details of a snapshot export to Amazon S3.

This data type is used as a response element in the DescribeExportTasks action.

See: newExportTask smart constructor.

Constructors

ExportTask'

Fields

totalExtractedDataInGB :: Maybe Int

The total amount of data exported, in gigabytes.

status :: Maybe Text

The progress status of the export task.

iamRoleArn :: Maybe Text

The name of the IAM role that is used to write to Amazon S3 when exporting a snapshot.

sourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the snapshot exported to Amazon S3.

exportOnly :: Maybe [Text]

The data exported from the snapshot. Valid values are the following:

database - Export all the data from a specified database.

database.table table-name - Export a table of the snapshot. This format is valid only for RDS for MySQL, RDS for MariaDB, and Aurora MySQL.

database.schema schema-name - Export a database schema of the snapshot. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

database.schema.table table-name - Export a table of the database schema. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

taskStartTime :: Maybe ISO8601

The time that the snapshot export task started.

warningMessage :: Maybe Text

A warning about the snapshot export task.

snapshotTime :: Maybe ISO8601

The time that the snapshot was created.

kmsKeyId :: Maybe Text

The key identifier of the Amazon Web Services KMS customer master key (CMK) that is used to encrypt the snapshot when it's exported to Amazon S3. The Amazon Web Services KMS CMK identifier is its key ARN, key ID, alias ARN, or alias name. The IAM role used for the snapshot export must have encryption and decryption permissions to use this Amazon Web Services KMS CMK.

taskEndTime :: Maybe ISO8601

The time that the snapshot export task completed.

exportTaskIdentifier :: Maybe Text

A unique identifier for the snapshot export task. This ID isn't an identifier for the Amazon S3 bucket where the snapshot is exported to.

s3Prefix :: Maybe Text

The Amazon S3 bucket prefix that is the file name and path of the exported snapshot.

percentProgress :: Maybe Int

The progress of the snapshot export task as a percentage.

s3Bucket :: Maybe Text

The Amazon S3 bucket that the snapshot is exported to.

failureCause :: Maybe Text

The reason the export failed, if it failed.

newExportTask :: ExportTask

Create a value of ExportTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:totalExtractedDataInGB:ExportTask', exportTask_totalExtractedDataInGB - The total amount of data exported, in gigabytes.

$sel:status:ExportTask', exportTask_status - The progress status of the export task.

$sel:iamRoleArn:ExportTask', exportTask_iamRoleArn - The name of the IAM role that is used to write to Amazon S3 when exporting a snapshot.

$sel:sourceArn:ExportTask', exportTask_sourceArn - The Amazon Resource Name (ARN) of the snapshot exported to Amazon S3.

$sel:exportOnly:ExportTask', exportTask_exportOnly - The data exported from the snapshot. Valid values are the following:

database - Export all the data from a specified database.

database.table table-name - Export a table of the snapshot. This format is valid only for RDS for MySQL, RDS for MariaDB, and Aurora MySQL.

database.schema schema-name - Export a database schema of the snapshot. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

database.schema.table table-name - Export a table of the database schema. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

$sel:taskStartTime:ExportTask', exportTask_taskStartTime - The time that the snapshot export task started.

$sel:warningMessage:ExportTask', exportTask_warningMessage - A warning about the snapshot export task.

$sel:snapshotTime:ExportTask', exportTask_snapshotTime - The time that the snapshot was created.

$sel:kmsKeyId:ExportTask', exportTask_kmsKeyId - The key identifier of the Amazon Web Services KMS customer master key (CMK) that is used to encrypt the snapshot when it's exported to Amazon S3. The Amazon Web Services KMS CMK identifier is its key ARN, key ID, alias ARN, or alias name. The IAM role used for the snapshot export must have encryption and decryption permissions to use this Amazon Web Services KMS CMK.

$sel:taskEndTime:ExportTask', exportTask_taskEndTime - The time that the snapshot export task completed.

$sel:exportTaskIdentifier:ExportTask', exportTask_exportTaskIdentifier - A unique identifier for the snapshot export task. This ID isn't an identifier for the Amazon S3 bucket where the snapshot is exported to.

$sel:s3Prefix:ExportTask', exportTask_s3Prefix - The Amazon S3 bucket prefix that is the file name and path of the exported snapshot.

$sel:percentProgress:ExportTask', exportTask_percentProgress - The progress of the snapshot export task as a percentage.

$sel:s3Bucket:ExportTask', exportTask_s3Bucket - The Amazon S3 bucket that the snapshot is exported to.

$sel:failureCause:ExportTask', exportTask_failureCause - The reason the export failed, if it failed.

Response Lenses

exportTask_totalExtractedDataInGB :: Lens' ExportTask (Maybe Int)

The total amount of data exported, in gigabytes.

exportTask_status :: Lens' ExportTask (Maybe Text)

The progress status of the export task.

exportTask_iamRoleArn :: Lens' ExportTask (Maybe Text)

The name of the IAM role that is used to write to Amazon S3 when exporting a snapshot.

exportTask_sourceArn :: Lens' ExportTask (Maybe Text)

The Amazon Resource Name (ARN) of the snapshot exported to Amazon S3.

exportTask_exportOnly :: Lens' ExportTask (Maybe [Text])

The data exported from the snapshot. Valid values are the following:

database - Export all the data from a specified database.

database.table table-name - Export a table of the snapshot. This format is valid only for RDS for MySQL, RDS for MariaDB, and Aurora MySQL.

database.schema schema-name - Export a database schema of the snapshot. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

database.schema.table table-name - Export a table of the database schema. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

exportTask_taskStartTime :: Lens' ExportTask (Maybe UTCTime)

The time that the snapshot export task started.

exportTask_warningMessage :: Lens' ExportTask (Maybe Text)

A warning about the snapshot export task.

exportTask_snapshotTime :: Lens' ExportTask (Maybe UTCTime)

The time that the snapshot was created.

exportTask_kmsKeyId :: Lens' ExportTask (Maybe Text)

The key identifier of the Amazon Web Services KMS customer master key (CMK) that is used to encrypt the snapshot when it's exported to Amazon S3. The Amazon Web Services KMS CMK identifier is its key ARN, key ID, alias ARN, or alias name. The IAM role used for the snapshot export must have encryption and decryption permissions to use this Amazon Web Services KMS CMK.

exportTask_taskEndTime :: Lens' ExportTask (Maybe UTCTime)

The time that the snapshot export task completed.

exportTask_exportTaskIdentifier :: Lens' ExportTask (Maybe Text)

A unique identifier for the snapshot export task. This ID isn't an identifier for the Amazon S3 bucket where the snapshot is exported to.

exportTask_s3Prefix :: Lens' ExportTask (Maybe Text)

The Amazon S3 bucket prefix that is the file name and path of the exported snapshot.

exportTask_percentProgress :: Lens' ExportTask (Maybe Int)

The progress of the snapshot export task as a percentage.

exportTask_s3Bucket :: Lens' ExportTask (Maybe Text)

The Amazon S3 bucket that the snapshot is exported to.

exportTask_failureCause :: Lens' ExportTask (Maybe Text)

The reason the export failed, if it failed.