Amazonka.RDS.StopDBInstanceAutomatedBackupsReplication

Description

Stops automated backup replication for a DB instance.

For more information, see Replicating Automated Backups to Another Amazon Web Services Region in the Amazon RDS User Guide.

Creating a Request

data StopDBInstanceAutomatedBackupsReplication

See: newStopDBInstanceAutomatedBackupsReplication smart constructor.

Constructors

StopDBInstanceAutomatedBackupsReplication'

Fields

sourceDBInstanceArn :: Text

The Amazon Resource Name (ARN) of the source DB instance for which to stop replicating automated backups, for example, arn:aws:rds:us-west-2:123456789012:db:mydatabase.

newStopDBInstanceAutomatedBackupsReplication

Arguments

:: Text
$sel:sourceDBInstanceArn:StopDBInstanceAutomatedBackupsReplication'

-> StopDBInstanceAutomatedBackupsReplication

Create a value of StopDBInstanceAutomatedBackupsReplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceDBInstanceArn:StopDBInstanceAutomatedBackupsReplication', stopDBInstanceAutomatedBackupsReplication_sourceDBInstanceArn - The Amazon Resource Name (ARN) of the source DB instance for which to stop replicating automated backups, for example, arn:aws:rds:us-west-2:123456789012:db:mydatabase.

Request Lenses

stopDBInstanceAutomatedBackupsReplication_sourceDBInstanceArn :: Lens' StopDBInstanceAutomatedBackupsReplication Text

The Amazon Resource Name (ARN) of the source DB instance for which to stop replicating automated backups, for example, arn:aws:rds:us-west-2:123456789012:db:mydatabase.

Destructuring the Response

data StopDBInstanceAutomatedBackupsReplicationResponse

See: newStopDBInstanceAutomatedBackupsReplicationResponse smart constructor.

Constructors

StopDBInstanceAutomatedBackupsReplicationResponse'

Fields

dbInstanceAutomatedBackup :: Maybe DBInstanceAutomatedBackup

httpStatus :: Int

The response's http status code.

newStopDBInstanceAutomatedBackupsReplicationResponse

Arguments

:: Int
$sel:httpStatus:StopDBInstanceAutomatedBackupsReplicationResponse'

-> StopDBInstanceAutomatedBackupsReplicationResponse

Create a value of StopDBInstanceAutomatedBackupsReplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstanceAutomatedBackup:StopDBInstanceAutomatedBackupsReplicationResponse', stopDBInstanceAutomatedBackupsReplicationResponse_dbInstanceAutomatedBackup - Undocumented member.

$sel:httpStatus:StopDBInstanceAutomatedBackupsReplicationResponse', stopDBInstanceAutomatedBackupsReplicationResponse_httpStatus - The response's http status code.

Response Lenses

stopDBInstanceAutomatedBackupsReplicationResponse_dbInstanceAutomatedBackup :: Lens' StopDBInstanceAutomatedBackupsReplicationResponse (Maybe DBInstanceAutomatedBackup)

Undocumented member.

stopDBInstanceAutomatedBackupsReplicationResponse_httpStatus :: Lens' StopDBInstanceAutomatedBackupsReplicationResponse Int

The response's http status code.