Amazonka.RDS.Types.DBClusterSnapshotAttribute

Description

Documentation

data DBClusterSnapshotAttribute

Contains the name and values of a manual DB cluster snapshot attribute.

Manual DB cluster snapshot attributes are used to authorize other Amazon Web Services accounts to restore a manual DB cluster snapshot. For more information, see the ModifyDBClusterSnapshotAttribute API action.

See: newDBClusterSnapshotAttribute smart constructor.

Constructors

DBClusterSnapshotAttribute'

Fields

attributeValues :: Maybe [Text]

The value(s) for the manual DB cluster snapshot attribute.

If the AttributeName field is set to restore, then this element returns a list of IDs of the Amazon Web Services accounts that are authorized to copy or restore the manual DB cluster snapshot. If a value of all is in the list, then the manual DB cluster snapshot is public and available for any Amazon Web Services account to copy or restore.

attributeName :: Maybe Text

The name of the manual DB cluster snapshot attribute.

The attribute named restore refers to the list of Amazon Web Services accounts that have permission to copy or restore the manual DB cluster snapshot. For more information, see the ModifyDBClusterSnapshotAttribute API action.

newDBClusterSnapshotAttribute :: DBClusterSnapshotAttribute

Create a value of DBClusterSnapshotAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeValues:DBClusterSnapshotAttribute', dbClusterSnapshotAttribute_attributeValues - The value(s) for the manual DB cluster snapshot attribute.

If the AttributeName field is set to restore, then this element returns a list of IDs of the Amazon Web Services accounts that are authorized to copy or restore the manual DB cluster snapshot. If a value of all is in the list, then the manual DB cluster snapshot is public and available for any Amazon Web Services account to copy or restore.

$sel:attributeName:DBClusterSnapshotAttribute', dbClusterSnapshotAttribute_attributeName - The name of the manual DB cluster snapshot attribute.

The attribute named restore refers to the list of Amazon Web Services accounts that have permission to copy or restore the manual DB cluster snapshot. For more information, see the ModifyDBClusterSnapshotAttribute API action.

dbClusterSnapshotAttribute_attributeValues :: Lens' DBClusterSnapshotAttribute (Maybe [Text])

The value(s) for the manual DB cluster snapshot attribute.

If the AttributeName field is set to restore, then this element returns a list of IDs of the Amazon Web Services accounts that are authorized to copy or restore the manual DB cluster snapshot. If a value of all is in the list, then the manual DB cluster snapshot is public and available for any Amazon Web Services account to copy or restore.

dbClusterSnapshotAttribute_attributeName :: Lens' DBClusterSnapshotAttribute (Maybe Text)

The name of the manual DB cluster snapshot attribute.

The attribute named restore refers to the list of Amazon Web Services accounts that have permission to copy or restore the manual DB cluster snapshot. For more information, see the ModifyDBClusterSnapshotAttribute API action.