Amazonka.RDS.Types.DBProxyEndpoint

Description

Documentation

data DBProxyEndpoint

The data structure representing an endpoint associated with a DB proxy. RDS automatically creates one endpoint for each DB proxy. For Aurora DB clusters, you can associate additional endpoints with the same DB proxy. These endpoints can be read/write or read-only. They can also reside in different VPCs than the associated DB proxy.

This data type is used as a response element in the DescribeDBProxyEndpoints operation.

See: newDBProxyEndpoint smart constructor.

Constructors

DBProxyEndpoint'

Fields

status :: Maybe DBProxyEndpointStatus

The current status of this DB proxy endpoint. A status of available means the endpoint is ready to handle requests. Other values indicate that you must wait for the endpoint to be ready, or take some action to resolve an issue.

dbProxyEndpointArn :: Maybe Text

The Amazon Resource Name (ARN) for the DB proxy endpoint.

targetRole :: Maybe DBProxyEndpointTargetRole

A value that indicates whether the DB proxy endpoint can be used for read/write or read-only operations.

vpcSubnetIds :: Maybe [Text]

The EC2 subnet IDs for the DB proxy endpoint.

vpcId :: Maybe Text

Provides the VPC ID of the DB proxy endpoint.

createdDate :: Maybe ISO8601

The date and time when the DB proxy endpoint was first created.

vpcSecurityGroupIds :: Maybe [Text]

Provides a list of VPC security groups that the DB proxy endpoint belongs to.

dbProxyName :: Maybe Text

The identifier for the DB proxy that is associated with this DB proxy endpoint.

endpoint :: Maybe Text

The endpoint that you can use to connect to the DB proxy. You include the endpoint value in the connection string for a database client application.

dbProxyEndpointName :: Maybe Text

The name for the DB proxy endpoint. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

isDefault :: Maybe Bool

A value that indicates whether this endpoint is the default endpoint for the associated DB proxy. Default DB proxy endpoints always have read/write capability. Other endpoints that you associate with the DB proxy can be either read/write or read-only.

newDBProxyEndpoint :: DBProxyEndpoint

Create a value of DBProxyEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DBProxyEndpoint', dbProxyEndpoint_status - The current status of this DB proxy endpoint. A status of available means the endpoint is ready to handle requests. Other values indicate that you must wait for the endpoint to be ready, or take some action to resolve an issue.

$sel:dbProxyEndpointArn:DBProxyEndpoint', dbProxyEndpoint_dbProxyEndpointArn - The Amazon Resource Name (ARN) for the DB proxy endpoint.

$sel:targetRole:DBProxyEndpoint', dbProxyEndpoint_targetRole - A value that indicates whether the DB proxy endpoint can be used for read/write or read-only operations.

$sel:vpcSubnetIds:DBProxyEndpoint', dbProxyEndpoint_vpcSubnetIds - The EC2 subnet IDs for the DB proxy endpoint.

$sel:vpcId:DBProxyEndpoint', dbProxyEndpoint_vpcId - Provides the VPC ID of the DB proxy endpoint.

$sel:createdDate:DBProxyEndpoint', dbProxyEndpoint_createdDate - The date and time when the DB proxy endpoint was first created.

$sel:vpcSecurityGroupIds:DBProxyEndpoint', dbProxyEndpoint_vpcSecurityGroupIds - Provides a list of VPC security groups that the DB proxy endpoint belongs to.

$sel:dbProxyName:DBProxyEndpoint', dbProxyEndpoint_dbProxyName - The identifier for the DB proxy that is associated with this DB proxy endpoint.

$sel:endpoint:DBProxyEndpoint', dbProxyEndpoint_endpoint - The endpoint that you can use to connect to the DB proxy. You include the endpoint value in the connection string for a database client application.

$sel:dbProxyEndpointName:DBProxyEndpoint', dbProxyEndpoint_dbProxyEndpointName - The name for the DB proxy endpoint. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

$sel:isDefault:DBProxyEndpoint', dbProxyEndpoint_isDefault - A value that indicates whether this endpoint is the default endpoint for the associated DB proxy. Default DB proxy endpoints always have read/write capability. Other endpoints that you associate with the DB proxy can be either read/write or read-only.

dbProxyEndpoint_status :: Lens' DBProxyEndpoint (Maybe DBProxyEndpointStatus)

The current status of this DB proxy endpoint. A status of available means the endpoint is ready to handle requests. Other values indicate that you must wait for the endpoint to be ready, or take some action to resolve an issue.

dbProxyEndpoint_dbProxyEndpointArn :: Lens' DBProxyEndpoint (Maybe Text)

The Amazon Resource Name (ARN) for the DB proxy endpoint.

dbProxyEndpoint_targetRole :: Lens' DBProxyEndpoint (Maybe DBProxyEndpointTargetRole)

A value that indicates whether the DB proxy endpoint can be used for read/write or read-only operations.

dbProxyEndpoint_vpcSubnetIds :: Lens' DBProxyEndpoint (Maybe [Text])

The EC2 subnet IDs for the DB proxy endpoint.

dbProxyEndpoint_vpcId :: Lens' DBProxyEndpoint (Maybe Text)

Provides the VPC ID of the DB proxy endpoint.

dbProxyEndpoint_createdDate :: Lens' DBProxyEndpoint (Maybe UTCTime)

The date and time when the DB proxy endpoint was first created.

dbProxyEndpoint_vpcSecurityGroupIds :: Lens' DBProxyEndpoint (Maybe [Text])

Provides a list of VPC security groups that the DB proxy endpoint belongs to.

dbProxyEndpoint_dbProxyName :: Lens' DBProxyEndpoint (Maybe Text)

The identifier for the DB proxy that is associated with this DB proxy endpoint.

dbProxyEndpoint_endpoint :: Lens' DBProxyEndpoint (Maybe Text)

The endpoint that you can use to connect to the DB proxy. You include the endpoint value in the connection string for a database client application.

dbProxyEndpoint_dbProxyEndpointName :: Lens' DBProxyEndpoint (Maybe Text)

The name for the DB proxy endpoint. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

dbProxyEndpoint_isDefault :: Lens' DBProxyEndpoint (Maybe Bool)

A value that indicates whether this endpoint is the default endpoint for the associated DB proxy. Default DB proxy endpoints always have read/write capability. Other endpoints that you associate with the DB proxy can be either read/write or read-only.