Amazonka.RDS.Types.DBSnapshot

Description

Documentation

data DBSnapshot

Contains the details of an Amazon RDS DB snapshot.

This data type is used as a response element in the DescribeDBSnapshots action.

See: newDBSnapshot smart constructor.

Constructors

DBSnapshot'

Fields

originalSnapshotCreateTime :: Maybe ISO8601

Specifies the time of the CreateDBSnapshot operation in Coordinated Universal Time (UTC). Doesn't change when the snapshot is copied.

engineVersion :: Maybe Text

Specifies the version of the database engine.

status :: Maybe Text

Specifies the status of this DB snapshot.

dbSnapshotArn :: Maybe Text

The Amazon Resource Name (ARN) for the DB snapshot.

masterUsername :: Maybe Text

Provides the master username for the DB snapshot.

sourceRegion :: Maybe Text

The Amazon Web Services Region that the DB snapshot was created in or copied from.

iAMDatabaseAuthenticationEnabled :: Maybe Bool

True if mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

iops :: Maybe Int

Specifies the Provisioned IOPS (I/O operations per second) value of the DB instance at the time of the snapshot.

vpcId :: Maybe Text

Provides the VPC ID associated with the DB snapshot.

instanceCreateTime :: Maybe ISO8601

Specifies the time in Coordinated Universal Time (UTC) when the DB instance, from which the snapshot was taken, was created.

tagList :: Maybe [Tag]

engine :: Maybe Text

Specifies the name of the database engine.

encrypted :: Maybe Bool

Specifies whether the DB snapshot is encrypted.

dbSnapshotIdentifier :: Maybe Text

Specifies the identifier for the DB snapshot.

processorFeatures :: Maybe [ProcessorFeature]

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance when the DB snapshot was created.

licenseModel :: Maybe Text

License model information for the restored DB instance.

sourceDBSnapshotIdentifier :: Maybe Text

The DB snapshot Amazon Resource Name (ARN) that the DB snapshot was copied from. It only has a value in the case of a cross-account or cross-Region copy.

snapshotType :: Maybe Text

Provides the type of the DB snapshot.

dbInstanceIdentifier :: Maybe Text

Specifies the DB instance identifier of the DB instance this DB snapshot was created from.

kmsKeyId :: Maybe Text

If Encrypted is true, the Amazon Web Services KMS key identifier for the encrypted DB snapshot.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

availabilityZone :: Maybe Text

Specifies the name of the Availability Zone the DB instance was located in at the time of the DB snapshot.

snapshotCreateTime :: Maybe ISO8601

Specifies when the snapshot was taken in Coordinated Universal Time (UTC). Changes for the copy when the snapshot is copied.

allocatedStorage :: Maybe Int

Specifies the allocated storage size in gibibytes (GiB).

dbiResourceId :: Maybe Text

The identifier for the source DB instance, which can't be changed and which is unique to an Amazon Web Services Region.

optionGroupName :: Maybe Text

Provides the option group name for the DB snapshot.

timezone :: Maybe Text

The time zone of the DB snapshot. In most cases, the Timezone element is empty. Timezone content appears only for snapshots taken from Microsoft SQL Server DB instances that were created with a time zone specified.

tdeCredentialArn :: Maybe Text

The ARN from the key store with which to associate the instance for TDE encryption.

percentProgress :: Maybe Int

The percentage of the estimated data that has been transferred.

port :: Maybe Int

Specifies the port that the database engine was listening on at the time of the snapshot.

storageType :: Maybe Text

Specifies the storage type associated with DB snapshot.

newDBSnapshot :: DBSnapshot

Create a value of DBSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:originalSnapshotCreateTime:DBSnapshot', dbSnapshot_originalSnapshotCreateTime - Specifies the time of the CreateDBSnapshot operation in Coordinated Universal Time (UTC). Doesn't change when the snapshot is copied.

$sel:engineVersion:DBSnapshot', dbSnapshot_engineVersion - Specifies the version of the database engine.

$sel:status:DBSnapshot', dbSnapshot_status - Specifies the status of this DB snapshot.

$sel:dbSnapshotArn:DBSnapshot', dbSnapshot_dbSnapshotArn - The Amazon Resource Name (ARN) for the DB snapshot.

$sel:masterUsername:DBSnapshot', dbSnapshot_masterUsername - Provides the master username for the DB snapshot.

$sel:sourceRegion:DBSnapshot', dbSnapshot_sourceRegion - The Amazon Web Services Region that the DB snapshot was created in or copied from.

$sel:iAMDatabaseAuthenticationEnabled:DBSnapshot', dbSnapshot_iAMDatabaseAuthenticationEnabled - True if mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

$sel:iops:DBSnapshot', dbSnapshot_iops - Specifies the Provisioned IOPS (I/O operations per second) value of the DB instance at the time of the snapshot.

$sel:vpcId:DBSnapshot', dbSnapshot_vpcId - Provides the VPC ID associated with the DB snapshot.

$sel:instanceCreateTime:DBSnapshot', dbSnapshot_instanceCreateTime - Specifies the time in Coordinated Universal Time (UTC) when the DB instance, from which the snapshot was taken, was created.

$sel:tagList:DBSnapshot', dbSnapshot_tagList - Undocumented member.

$sel:engine:DBSnapshot', dbSnapshot_engine - Specifies the name of the database engine.

$sel:encrypted:DBSnapshot', dbSnapshot_encrypted - Specifies whether the DB snapshot is encrypted.

$sel:dbSnapshotIdentifier:DBSnapshot', dbSnapshot_dbSnapshotIdentifier - Specifies the identifier for the DB snapshot.

$sel:processorFeatures:DBSnapshot', dbSnapshot_processorFeatures - The number of CPU cores and the number of threads per core for the DB instance class of the DB instance when the DB snapshot was created.

$sel:licenseModel:DBSnapshot', dbSnapshot_licenseModel - License model information for the restored DB instance.

$sel:sourceDBSnapshotIdentifier:DBSnapshot', dbSnapshot_sourceDBSnapshotIdentifier - The DB snapshot Amazon Resource Name (ARN) that the DB snapshot was copied from. It only has a value in the case of a cross-account or cross-Region copy.

$sel:snapshotType:DBSnapshot', dbSnapshot_snapshotType - Provides the type of the DB snapshot.

$sel:dbInstanceIdentifier:DBSnapshot', dbSnapshot_dbInstanceIdentifier - Specifies the DB instance identifier of the DB instance this DB snapshot was created from.

$sel:kmsKeyId:DBSnapshot', dbSnapshot_kmsKeyId - If Encrypted is true, the Amazon Web Services KMS key identifier for the encrypted DB snapshot.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

$sel:availabilityZone:DBSnapshot', dbSnapshot_availabilityZone - Specifies the name of the Availability Zone the DB instance was located in at the time of the DB snapshot.

$sel:snapshotCreateTime:DBSnapshot', dbSnapshot_snapshotCreateTime - Specifies when the snapshot was taken in Coordinated Universal Time (UTC). Changes for the copy when the snapshot is copied.

$sel:allocatedStorage:DBSnapshot', dbSnapshot_allocatedStorage - Specifies the allocated storage size in gibibytes (GiB).

$sel:dbiResourceId:DBSnapshot', dbSnapshot_dbiResourceId - The identifier for the source DB instance, which can't be changed and which is unique to an Amazon Web Services Region.

$sel:optionGroupName:DBSnapshot', dbSnapshot_optionGroupName - Provides the option group name for the DB snapshot.

$sel:timezone:DBSnapshot', dbSnapshot_timezone - The time zone of the DB snapshot. In most cases, the Timezone element is empty. Timezone content appears only for snapshots taken from Microsoft SQL Server DB instances that were created with a time zone specified.

$sel:tdeCredentialArn:DBSnapshot', dbSnapshot_tdeCredentialArn - The ARN from the key store with which to associate the instance for TDE encryption.

$sel:percentProgress:DBSnapshot', dbSnapshot_percentProgress - The percentage of the estimated data that has been transferred.

$sel:port:DBSnapshot', dbSnapshot_port - Specifies the port that the database engine was listening on at the time of the snapshot.

$sel:storageType:DBSnapshot', dbSnapshot_storageType - Specifies the storage type associated with DB snapshot.

dbSnapshot_originalSnapshotCreateTime :: Lens' DBSnapshot (Maybe UTCTime)

Specifies the time of the CreateDBSnapshot operation in Coordinated Universal Time (UTC). Doesn't change when the snapshot is copied.

dbSnapshot_engineVersion :: Lens' DBSnapshot (Maybe Text)

Specifies the version of the database engine.

dbSnapshot_status :: Lens' DBSnapshot (Maybe Text)

Specifies the status of this DB snapshot.

dbSnapshot_dbSnapshotArn :: Lens' DBSnapshot (Maybe Text)

The Amazon Resource Name (ARN) for the DB snapshot.

dbSnapshot_masterUsername :: Lens' DBSnapshot (Maybe Text)

Provides the master username for the DB snapshot.

dbSnapshot_sourceRegion :: Lens' DBSnapshot (Maybe Text)

The Amazon Web Services Region that the DB snapshot was created in or copied from.

dbSnapshot_iAMDatabaseAuthenticationEnabled :: Lens' DBSnapshot (Maybe Bool)

True if mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

dbSnapshot_iops :: Lens' DBSnapshot (Maybe Int)

Specifies the Provisioned IOPS (I/O operations per second) value of the DB instance at the time of the snapshot.

dbSnapshot_vpcId :: Lens' DBSnapshot (Maybe Text)

Provides the VPC ID associated with the DB snapshot.

dbSnapshot_instanceCreateTime :: Lens' DBSnapshot (Maybe UTCTime)

Specifies the time in Coordinated Universal Time (UTC) when the DB instance, from which the snapshot was taken, was created.

dbSnapshot_tagList :: Lens' DBSnapshot (Maybe [Tag])

Undocumented member.

dbSnapshot_engine :: Lens' DBSnapshot (Maybe Text)

Specifies the name of the database engine.

dbSnapshot_encrypted :: Lens' DBSnapshot (Maybe Bool)

Specifies whether the DB snapshot is encrypted.

dbSnapshot_dbSnapshotIdentifier :: Lens' DBSnapshot (Maybe Text)

Specifies the identifier for the DB snapshot.

dbSnapshot_processorFeatures :: Lens' DBSnapshot (Maybe [ProcessorFeature])

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance when the DB snapshot was created.

dbSnapshot_licenseModel :: Lens' DBSnapshot (Maybe Text)

License model information for the restored DB instance.

dbSnapshot_sourceDBSnapshotIdentifier :: Lens' DBSnapshot (Maybe Text)

The DB snapshot Amazon Resource Name (ARN) that the DB snapshot was copied from. It only has a value in the case of a cross-account or cross-Region copy.

dbSnapshot_snapshotType :: Lens' DBSnapshot (Maybe Text)

Provides the type of the DB snapshot.

dbSnapshot_dbInstanceIdentifier :: Lens' DBSnapshot (Maybe Text)

Specifies the DB instance identifier of the DB instance this DB snapshot was created from.

dbSnapshot_kmsKeyId :: Lens' DBSnapshot (Maybe Text)

If Encrypted is true, the Amazon Web Services KMS key identifier for the encrypted DB snapshot.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

dbSnapshot_availabilityZone :: Lens' DBSnapshot (Maybe Text)

Specifies the name of the Availability Zone the DB instance was located in at the time of the DB snapshot.

dbSnapshot_snapshotCreateTime :: Lens' DBSnapshot (Maybe UTCTime)

Specifies when the snapshot was taken in Coordinated Universal Time (UTC). Changes for the copy when the snapshot is copied.

dbSnapshot_allocatedStorage :: Lens' DBSnapshot (Maybe Int)

Specifies the allocated storage size in gibibytes (GiB).

dbSnapshot_dbiResourceId :: Lens' DBSnapshot (Maybe Text)

The identifier for the source DB instance, which can't be changed and which is unique to an Amazon Web Services Region.

dbSnapshot_optionGroupName :: Lens' DBSnapshot (Maybe Text)

Provides the option group name for the DB snapshot.

dbSnapshot_timezone :: Lens' DBSnapshot (Maybe Text)

The time zone of the DB snapshot. In most cases, the Timezone element is empty. Timezone content appears only for snapshots taken from Microsoft SQL Server DB instances that were created with a time zone specified.

dbSnapshot_tdeCredentialArn :: Lens' DBSnapshot (Maybe Text)

The ARN from the key store with which to associate the instance for TDE encryption.

dbSnapshot_percentProgress :: Lens' DBSnapshot (Maybe Int)

The percentage of the estimated data that has been transferred.

dbSnapshot_port :: Lens' DBSnapshot (Maybe Int)

Specifies the port that the database engine was listening on at the time of the snapshot.

dbSnapshot_storageType :: Lens' DBSnapshot (Maybe Text)

Specifies the storage type associated with DB snapshot.