Amazonka.RDS.Types.DomainMembership

Description

Documentation

data DomainMembership

An Active Directory Domain membership record associated with the DB instance or cluster.

See: newDomainMembership smart constructor.

Constructors

DomainMembership'

Fields

status :: Maybe Text

The status of the Active Directory Domain membership for the DB instance or cluster. Values include joined, pending-join, failed, and so on.

fqdn :: Maybe Text

The fully qualified domain name of the Active Directory Domain.

domain :: Maybe Text

The identifier of the Active Directory Domain.

iAMRoleName :: Maybe Text

The name of the IAM role to be used when making API calls to the Directory Service.

newDomainMembership :: DomainMembership

Create a value of DomainMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DomainMembership', domainMembership_status - The status of the Active Directory Domain membership for the DB instance or cluster. Values include joined, pending-join, failed, and so on.

$sel:fqdn:DomainMembership', domainMembership_fqdn - The fully qualified domain name of the Active Directory Domain.

$sel:domain:DomainMembership', domainMembership_domain - The identifier of the Active Directory Domain.

$sel:iAMRoleName:DomainMembership', domainMembership_iAMRoleName - The name of the IAM role to be used when making API calls to the Directory Service.

domainMembership_status :: Lens' DomainMembership (Maybe Text)

The status of the Active Directory Domain membership for the DB instance or cluster. Values include joined, pending-join, failed, and so on.

domainMembership_fqdn :: Lens' DomainMembership (Maybe Text)

The fully qualified domain name of the Active Directory Domain.

domainMembership_domain :: Lens' DomainMembership (Maybe Text)

The identifier of the Active Directory Domain.

domainMembership_iAMRoleName :: Lens' DomainMembership (Maybe Text)

The name of the IAM role to be used when making API calls to the Directory Service.