Amazonka.RDS.Types.OptionConfiguration

Description

Documentation

data OptionConfiguration

A list of all available options

See: newOptionConfiguration smart constructor.

Constructors

OptionConfiguration'

Fields

optionSettings :: Maybe [OptionSetting]

The option settings to include in an option group.

vpcSecurityGroupMemberships :: Maybe [Text]

A list of VpcSecurityGroupMembership name strings used for this option.

dbSecurityGroupMemberships :: Maybe [Text]

A list of DBSecurityGroupMembership name strings used for this option.

optionVersion :: Maybe Text

The version for the option.

port :: Maybe Int

The optional port for the option.

optionName :: Text

The configuration of options to include in a group.

newOptionConfiguration

Arguments

:: Text
$sel:optionName:OptionConfiguration'

-> OptionConfiguration

Create a value of OptionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionSettings:OptionConfiguration', optionConfiguration_optionSettings - The option settings to include in an option group.

$sel:vpcSecurityGroupMemberships:OptionConfiguration', optionConfiguration_vpcSecurityGroupMemberships - A list of VpcSecurityGroupMembership name strings used for this option.

$sel:dbSecurityGroupMemberships:OptionConfiguration', optionConfiguration_dbSecurityGroupMemberships - A list of DBSecurityGroupMembership name strings used for this option.

$sel:optionVersion:OptionConfiguration', optionConfiguration_optionVersion - The version for the option.

$sel:port:OptionConfiguration', optionConfiguration_port - The optional port for the option.

$sel:optionName:OptionConfiguration', optionConfiguration_optionName - The configuration of options to include in a group.

optionConfiguration_optionSettings :: Lens' OptionConfiguration (Maybe [OptionSetting])

The option settings to include in an option group.

optionConfiguration_vpcSecurityGroupMemberships :: Lens' OptionConfiguration (Maybe [Text])

A list of VpcSecurityGroupMembership name strings used for this option.

optionConfiguration_dbSecurityGroupMemberships :: Lens' OptionConfiguration (Maybe [Text])

A list of DBSecurityGroupMembership name strings used for this option.

optionConfiguration_optionVersion :: Lens' OptionConfiguration (Maybe Text)

The version for the option.

optionConfiguration_port :: Lens' OptionConfiguration (Maybe Int)

The optional port for the option.

optionConfiguration_optionName :: Lens' OptionConfiguration Text

The configuration of options to include in a group.