Amazonka.RDS.Types.OptionSetting

Description

Documentation

data OptionSetting

Option settings are the actual settings being applied or configured for that option. It is used when you modify an option group or describe option groups. For example, the NATIVE_NETWORK_ENCRYPTION option has a setting called SQLNET.ENCRYPTION_SERVER that can have several different values.

See: newOptionSetting smart constructor.

Constructors

OptionSetting'

Fields

isCollection :: Maybe Bool

Indicates if the option setting is part of a collection.

applyType :: Maybe Text

The DB engine specific parameter type.

value :: Maybe Text

The current value of the option setting.

name :: Maybe Text

The name of the option that has settings that you can set.

defaultValue :: Maybe Text

The default value of the option setting.

isModifiable :: Maybe Bool

A Boolean value that, when true, indicates the option setting can be modified from the default.

dataType :: Maybe Text

The data type of the option setting.

allowedValues :: Maybe Text

The allowed values of the option setting.

description :: Maybe Text

The description of the option setting.

newOptionSetting :: OptionSetting

Create a value of OptionSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isCollection:OptionSetting', optionSetting_isCollection - Indicates if the option setting is part of a collection.

$sel:applyType:OptionSetting', optionSetting_applyType - The DB engine specific parameter type.

$sel:value:OptionSetting', optionSetting_value - The current value of the option setting.

$sel:name:OptionSetting', optionSetting_name - The name of the option that has settings that you can set.

$sel:defaultValue:OptionSetting', optionSetting_defaultValue - The default value of the option setting.

$sel:isModifiable:OptionSetting', optionSetting_isModifiable - A Boolean value that, when true, indicates the option setting can be modified from the default.

$sel:dataType:OptionSetting', optionSetting_dataType - The data type of the option setting.

$sel:allowedValues:OptionSetting', optionSetting_allowedValues - The allowed values of the option setting.

$sel:description:OptionSetting', optionSetting_description - The description of the option setting.

optionSetting_isCollection :: Lens' OptionSetting (Maybe Bool)

Indicates if the option setting is part of a collection.

optionSetting_applyType :: Lens' OptionSetting (Maybe Text)

The DB engine specific parameter type.

optionSetting_value :: Lens' OptionSetting (Maybe Text)

The current value of the option setting.

optionSetting_name :: Lens' OptionSetting (Maybe Text)

The name of the option that has settings that you can set.

optionSetting_defaultValue :: Lens' OptionSetting (Maybe Text)

The default value of the option setting.

optionSetting_isModifiable :: Lens' OptionSetting (Maybe Bool)

A Boolean value that, when true, indicates the option setting can be modified from the default.

optionSetting_dataType :: Lens' OptionSetting (Maybe Text)

The data type of the option setting.

optionSetting_allowedValues :: Lens' OptionSetting (Maybe Text)

The allowed values of the option setting.

optionSetting_description :: Lens' OptionSetting (Maybe Text)

The description of the option setting.