Amazonka.RDS.Types.ScalingConfigurationInfo

Description

Documentation

data ScalingConfigurationInfo

Shows the scaling configuration for an Aurora DB cluster in serverless DB engine mode.

For more information, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

See: newScalingConfigurationInfo smart constructor.

Constructors

ScalingConfigurationInfo'

Fields

secondsUntilAutoPause :: Maybe Int

The remaining amount of time, in seconds, before the Aurora DB cluster in serverless mode is paused. A DB cluster can be paused only when it's idle (it has no connections).

timeoutAction :: Maybe Text

The action that occurs when Aurora times out while attempting to change the capacity of an Aurora Serverless cluster. The value is either ForceApplyCapacityChange or RollbackCapacityChange.

ForceApplyCapacityChange, the default, sets the capacity to the specified value as soon as possible.

RollbackCapacityChange ignores the capacity change if a scaling point isn't found in the timeout period.

autoPause :: Maybe Bool

A value that indicates whether automatic pause is allowed for the Aurora DB cluster in serverless DB engine mode.

When the value is set to false for an Aurora Serverless DB cluster, the DB cluster automatically resumes.

maxCapacity :: Maybe Int

The maximum capacity for an Aurora DB cluster in serverless DB engine mode.

minCapacity :: Maybe Int

The maximum capacity for the Aurora DB cluster in serverless DB engine mode.

secondsBeforeTimeout :: Maybe Int

The number of seconds before scaling times out. What happens when an attempted scaling action times out is determined by the TimeoutAction setting.

newScalingConfigurationInfo :: ScalingConfigurationInfo

Create a value of ScalingConfigurationInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secondsUntilAutoPause:ScalingConfigurationInfo', scalingConfigurationInfo_secondsUntilAutoPause - The remaining amount of time, in seconds, before the Aurora DB cluster in serverless mode is paused. A DB cluster can be paused only when it's idle (it has no connections).

$sel:timeoutAction:ScalingConfigurationInfo', scalingConfigurationInfo_timeoutAction - The action that occurs when Aurora times out while attempting to change the capacity of an Aurora Serverless cluster. The value is either ForceApplyCapacityChange or RollbackCapacityChange.

ForceApplyCapacityChange, the default, sets the capacity to the specified value as soon as possible.

RollbackCapacityChange ignores the capacity change if a scaling point isn't found in the timeout period.

$sel:autoPause:ScalingConfigurationInfo', scalingConfigurationInfo_autoPause - A value that indicates whether automatic pause is allowed for the Aurora DB cluster in serverless DB engine mode.

When the value is set to false for an Aurora Serverless DB cluster, the DB cluster automatically resumes.

$sel:maxCapacity:ScalingConfigurationInfo', scalingConfigurationInfo_maxCapacity - The maximum capacity for an Aurora DB cluster in serverless DB engine mode.

$sel:minCapacity:ScalingConfigurationInfo', scalingConfigurationInfo_minCapacity - The maximum capacity for the Aurora DB cluster in serverless DB engine mode.

$sel:secondsBeforeTimeout:ScalingConfigurationInfo', scalingConfigurationInfo_secondsBeforeTimeout - The number of seconds before scaling times out. What happens when an attempted scaling action times out is determined by the TimeoutAction setting.

scalingConfigurationInfo_secondsUntilAutoPause :: Lens' ScalingConfigurationInfo (Maybe Int)

The remaining amount of time, in seconds, before the Aurora DB cluster in serverless mode is paused. A DB cluster can be paused only when it's idle (it has no connections).

scalingConfigurationInfo_timeoutAction :: Lens' ScalingConfigurationInfo (Maybe Text)

The action that occurs when Aurora times out while attempting to change the capacity of an Aurora Serverless cluster. The value is either ForceApplyCapacityChange or RollbackCapacityChange.

ForceApplyCapacityChange, the default, sets the capacity to the specified value as soon as possible.

RollbackCapacityChange ignores the capacity change if a scaling point isn't found in the timeout period.

scalingConfigurationInfo_autoPause :: Lens' ScalingConfigurationInfo (Maybe Bool)

A value that indicates whether automatic pause is allowed for the Aurora DB cluster in serverless DB engine mode.

When the value is set to false for an Aurora Serverless DB cluster, the DB cluster automatically resumes.

scalingConfigurationInfo_maxCapacity :: Lens' ScalingConfigurationInfo (Maybe Int)

The maximum capacity for an Aurora DB cluster in serverless DB engine mode.

scalingConfigurationInfo_minCapacity :: Lens' ScalingConfigurationInfo (Maybe Int)

The maximum capacity for the Aurora DB cluster in serverless DB engine mode.

scalingConfigurationInfo_secondsBeforeTimeout :: Lens' ScalingConfigurationInfo (Maybe Int)

The number of seconds before scaling times out. What happens when an attempted scaling action times out is determined by the TimeoutAction setting.