Amazonka.RDS.Types.ValidStorageOptions

Description

Documentation

data ValidStorageOptions

Information about valid modifications that you can make to your DB instance. Contains the result of a successful call to the DescribeValidDBInstanceModifications action.

See: newValidStorageOptions smart constructor.

Constructors

ValidStorageOptions'

Fields

storageSize :: Maybe [Range]

The valid range of storage in gibibytes (GiB). For example, 100 to 16384.

provisionedIops :: Maybe [Range]

The valid range of provisioned IOPS. For example, 1000-20000.

iopsToStorageRatio :: Maybe [DoubleRange]

The valid range of Provisioned IOPS to gibibytes of storage multiplier. For example, 3-10, which means that provisioned IOPS can be between 3 and 10 times storage.

supportsStorageAutoscaling :: Maybe Bool

Whether or not Amazon RDS can automatically scale storage for DB instances that use the new instance class.

storageType :: Maybe Text

The valid storage types for your DB instance. For example, gp2, io1.

newValidStorageOptions :: ValidStorageOptions

Create a value of ValidStorageOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:storageSize:ValidStorageOptions', validStorageOptions_storageSize - The valid range of storage in gibibytes (GiB). For example, 100 to 16384.

$sel:provisionedIops:ValidStorageOptions', validStorageOptions_provisionedIops - The valid range of provisioned IOPS. For example, 1000-20000.

$sel:iopsToStorageRatio:ValidStorageOptions', validStorageOptions_iopsToStorageRatio - The valid range of Provisioned IOPS to gibibytes of storage multiplier. For example, 3-10, which means that provisioned IOPS can be between 3 and 10 times storage.

$sel:supportsStorageAutoscaling:ValidStorageOptions', validStorageOptions_supportsStorageAutoscaling - Whether or not Amazon RDS can automatically scale storage for DB instances that use the new instance class.

$sel:storageType:ValidStorageOptions', validStorageOptions_storageType - The valid storage types for your DB instance. For example, gp2, io1.

validStorageOptions_storageSize :: Lens' ValidStorageOptions (Maybe [Range])

The valid range of storage in gibibytes (GiB). For example, 100 to 16384.

validStorageOptions_provisionedIops :: Lens' ValidStorageOptions (Maybe [Range])

The valid range of provisioned IOPS. For example, 1000-20000.

validStorageOptions_iopsToStorageRatio :: Lens' ValidStorageOptions (Maybe [DoubleRange])

The valid range of Provisioned IOPS to gibibytes of storage multiplier. For example, 3-10, which means that provisioned IOPS can be between 3 and 10 times storage.

validStorageOptions_supportsStorageAutoscaling :: Lens' ValidStorageOptions (Maybe Bool)

Whether or not Amazon RDS can automatically scale storage for DB instances that use the new instance class.

validStorageOptions_storageType :: Lens' ValidStorageOptions (Maybe Text)

The valid storage types for your DB instance. For example, gp2, io1.