Amazonka.RDSData.BeginTransaction

Description

Starts a SQL transaction.

 <important> <p>A transaction can run for a maximum of 24 hours. A transaction is terminated and rolled back automatically after 24 hours.</p> <p>A transaction times out if no calls use its transaction ID in three minutes. If a transaction times out before it's committed, it's rolled back automatically.</p> <p>DDL statements inside a transaction cause an implicit commit. We recommend that you run each DDL statement in a separate <code>ExecuteStatement</code> call with <code>continueAfterTimeout</code> enabled.</p> </important>

Creating a Request

data BeginTransaction

The request parameters represent the input of a request to start a SQL transaction.

See: newBeginTransaction smart constructor.

Constructors

BeginTransaction'

Fields

database :: Maybe Text

The name of the database.

schema :: Maybe Text

The name of the database schema.

resourceArn :: Text

The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

secretArn :: Text

The name or ARN of the secret that enables access to the DB cluster.

newBeginTransaction

Arguments

:: Text
$sel:resourceArn:BeginTransaction'

-> Text
$sel:secretArn:BeginTransaction'

-> BeginTransaction

Create a value of BeginTransaction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:database:BeginTransaction', beginTransaction_database - The name of the database.

$sel:schema:BeginTransaction', beginTransaction_schema - The name of the database schema.

$sel:resourceArn:BeginTransaction', beginTransaction_resourceArn - The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

$sel:secretArn:BeginTransaction', beginTransaction_secretArn - The name or ARN of the secret that enables access to the DB cluster.

Request Lenses

beginTransaction_database :: Lens' BeginTransaction (Maybe Text)

The name of the database.

beginTransaction_schema :: Lens' BeginTransaction (Maybe Text)

The name of the database schema.

beginTransaction_resourceArn :: Lens' BeginTransaction Text

The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

beginTransaction_secretArn :: Lens' BeginTransaction Text

The name or ARN of the secret that enables access to the DB cluster.

Destructuring the Response

data BeginTransactionResponse

The response elements represent the output of a request to start a SQL transaction.

See: newBeginTransactionResponse smart constructor.

Constructors

BeginTransactionResponse'

Fields

transactionId :: Maybe Text

The transaction ID of the transaction started by the call.

httpStatus :: Int

The response's http status code.

newBeginTransactionResponse

Arguments

:: Int
$sel:httpStatus:BeginTransactionResponse'

-> BeginTransactionResponse

Create a value of BeginTransactionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transactionId:BeginTransactionResponse', beginTransactionResponse_transactionId - The transaction ID of the transaction started by the call.

$sel:httpStatus:BeginTransactionResponse', beginTransactionResponse_httpStatus - The response's http status code.

Response Lenses

beginTransactionResponse_transactionId :: Lens' BeginTransactionResponse (Maybe Text)

The transaction ID of the transaction started by the call.

beginTransactionResponse_httpStatus :: Lens' BeginTransactionResponse Int

The response's http status code.