Amazonka.RDSData.Lens

Description

Operations

RollbackTransaction

rollbackTransaction_resourceArn :: Lens' RollbackTransaction Text

The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

rollbackTransaction_secretArn :: Lens' RollbackTransaction Text

The name or ARN of the secret that enables access to the DB cluster.

rollbackTransaction_transactionId :: Lens' RollbackTransaction Text

The identifier of the transaction to roll back.

rollbackTransactionResponse_transactionStatus :: Lens' RollbackTransactionResponse (Maybe Text)

The status of the rollback operation.

rollbackTransactionResponse_httpStatus :: Lens' RollbackTransactionResponse Int

The response's http status code.

BeginTransaction

beginTransaction_database :: Lens' BeginTransaction (Maybe Text)

The name of the database.

beginTransaction_schema :: Lens' BeginTransaction (Maybe Text)

The name of the database schema.

beginTransaction_resourceArn :: Lens' BeginTransaction Text

The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

beginTransaction_secretArn :: Lens' BeginTransaction Text

The name or ARN of the secret that enables access to the DB cluster.

beginTransactionResponse_transactionId :: Lens' BeginTransactionResponse (Maybe Text)

The transaction ID of the transaction started by the call.

beginTransactionResponse_httpStatus :: Lens' BeginTransactionResponse Int

The response's http status code.

BatchExecuteStatement

batchExecuteStatement_database :: Lens' BatchExecuteStatement (Maybe Text)

The name of the database.

batchExecuteStatement_parameterSets :: Lens' BatchExecuteStatement (Maybe [[SqlParameter]])

The parameter set for the batch operation.

The SQL statement is executed as many times as the number of parameter sets provided. To execute a SQL statement with no parameters, use one of the following options:

Specify one or more empty parameter sets.

Use the ExecuteStatement operation instead of the BatchExecuteStatement operation.

Array parameters are not supported.

batchExecuteStatement_transactionId :: Lens' BatchExecuteStatement (Maybe Text)

The identifier of a transaction that was started by using the BeginTransaction operation. Specify the transaction ID of the transaction that you want to include the SQL statement in.

If the SQL statement is not part of a transaction, don't set this parameter.

batchExecuteStatement_schema :: Lens' BatchExecuteStatement (Maybe Text)

The name of the database schema.

batchExecuteStatement_resourceArn :: Lens' BatchExecuteStatement Text

The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

batchExecuteStatement_secretArn :: Lens' BatchExecuteStatement Text

The name or ARN of the secret that enables access to the DB cluster.

batchExecuteStatement_sql :: Lens' BatchExecuteStatement Text

The SQL statement to run.

batchExecuteStatementResponse_updateResults :: Lens' BatchExecuteStatementResponse (Maybe [UpdateResult])

The execution results of each batch entry.

batchExecuteStatementResponse_httpStatus :: Lens' BatchExecuteStatementResponse Int

The response's http status code.

ExecuteStatement

executeStatement_database :: Lens' ExecuteStatement (Maybe Text)

The name of the database.

executeStatement_transactionId :: Lens' ExecuteStatement (Maybe Text)

The identifier of a transaction that was started by using the BeginTransaction operation. Specify the transaction ID of the transaction that you want to include the SQL statement in.

If the SQL statement is not part of a transaction, don't set this parameter.

executeStatement_schema :: Lens' ExecuteStatement (Maybe Text)

The name of the database schema.

Currently, the schema parameter isn't supported.

executeStatement_parameters :: Lens' ExecuteStatement (Maybe [SqlParameter])

The parameters for the SQL statement.

Array parameters are not supported.

executeStatement_includeResultMetadata :: Lens' ExecuteStatement (Maybe Bool)

A value that indicates whether to include metadata in the results.

executeStatement_resultSetOptions :: Lens' ExecuteStatement (Maybe ResultSetOptions)

Options that control how the result set is returned.

executeStatement_continueAfterTimeout :: Lens' ExecuteStatement (Maybe Bool)

A value that indicates whether to continue running the statement after the call times out. By default, the statement stops running when the call times out.

For DDL statements, we recommend continuing to run the statement after the call times out. When a DDL statement terminates before it is finished running, it can result in errors and possibly corrupted data structures.

executeStatement_resourceArn :: Lens' ExecuteStatement Text

The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

executeStatement_secretArn :: Lens' ExecuteStatement Text

The name or ARN of the secret that enables access to the DB cluster.

executeStatement_sql :: Lens' ExecuteStatement Text

The SQL statement to run.

executeStatementResponse_records :: Lens' ExecuteStatementResponse (Maybe [[Field]])

The records returned by the SQL statement.

executeStatementResponse_columnMetadata :: Lens' ExecuteStatementResponse (Maybe [ColumnMetadata])

Metadata for the columns included in the results.

executeStatementResponse_generatedFields :: Lens' ExecuteStatementResponse (Maybe [Field])

Values for fields generated during the request.

 <note> <p>The <code>generatedFields</code> data isn't supported by Aurora PostgreSQL. To get the values of generated fields, use the <code>RETURNING</code> clause. For more information, see <a href="https://www.postgresql.org/docs/10/dml-returning.html">Returning Data From Modified Rows</a> in the PostgreSQL documentation.</p> </note>

executeStatementResponse_numberOfRecordsUpdated :: Lens' ExecuteStatementResponse (Maybe Integer)

The number of records updated by the request.

executeStatementResponse_httpStatus :: Lens' ExecuteStatementResponse Int

The response's http status code.

CommitTransaction

commitTransaction_resourceArn :: Lens' CommitTransaction Text

The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

commitTransaction_secretArn :: Lens' CommitTransaction Text

The name or ARN of the secret that enables access to the DB cluster.

commitTransaction_transactionId :: Lens' CommitTransaction Text

The identifier of the transaction to end and commit.

commitTransactionResponse_transactionStatus :: Lens' CommitTransactionResponse (Maybe Text)

The status of the commit operation.

commitTransactionResponse_httpStatus :: Lens' CommitTransactionResponse Int

The response's http status code.

Types

ArrayValue

arrayValue_longValues :: Lens' ArrayValue (Maybe [Integer])

An array of floating point numbers.

arrayValue_doubleValues :: Lens' ArrayValue (Maybe [Double])

An array of integers.

arrayValue_stringValues :: Lens' ArrayValue (Maybe [Text])

An array of strings.

arrayValue_arrayValues :: Lens' ArrayValue (Maybe [ArrayValue])

An array of arrays.

arrayValue_booleanValues :: Lens' ArrayValue (Maybe [Bool])

An array of Boolean values.

ColumnMetadata

columnMetadata_typeName :: Lens' ColumnMetadata (Maybe Text)

The database-specific data type of the column.

columnMetadata_isCaseSensitive :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether the column is case-sensitive.

columnMetadata_isCurrency :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether the column contains currency values.

columnMetadata_scale :: Lens' ColumnMetadata (Maybe Int)

The scale value of a decimal number column.

columnMetadata_precision :: Lens' ColumnMetadata (Maybe Int)

The precision value of a decimal number column.

columnMetadata_schemaName :: Lens' ColumnMetadata (Maybe Text)

The name of the schema that owns the table that includes the column.

columnMetadata_isAutoIncrement :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether the column increments automatically.

columnMetadata_name :: Lens' ColumnMetadata (Maybe Text)

The name of the column.

columnMetadata_arrayBaseColumnType :: Lens' ColumnMetadata (Maybe Int)

The type of the column.

columnMetadata_type :: Lens' ColumnMetadata (Maybe Int)

The type of the column.

columnMetadata_isSigned :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether an integer column is signed.

columnMetadata_label :: Lens' ColumnMetadata (Maybe Text)

The label for the column.

columnMetadata_nullable :: Lens' ColumnMetadata (Maybe Int)

A value that indicates whether the column is nullable.

columnMetadata_tableName :: Lens' ColumnMetadata (Maybe Text)

The name of the table that includes the column.

Field

field_doubleValue :: Lens' Field (Maybe Double)

A value of double data type.

field_stringValue :: Lens' Field (Maybe Text)

A value of string data type.

field_longValue :: Lens' Field (Maybe Integer)

A value of long data type.

field_booleanValue :: Lens' Field (Maybe Bool)

A value of Boolean data type.

field_arrayValue :: Lens' Field (Maybe ArrayValue)

An array of values.

field_blobValue :: Lens' Field (Maybe ByteString)

A value of BLOB data type.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

field_isNull :: Lens' Field (Maybe Bool)

A NULL value.

ResultSetOptions

resultSetOptions_decimalReturnType :: Lens' ResultSetOptions (Maybe DecimalReturnType)

A value that indicates how a field of DECIMAL type is represented in the response. The value of STRING, the default, specifies that it is converted to a String value. The value of DOUBLE_OR_LONG specifies that it is converted to a Long value if its scale is 0, or to a Double value otherwise.

Conversion to Double or Long can result in roundoff errors due to precision loss. We recommend converting to String, especially when working with currency values.

SqlParameter

sqlParameter_value :: Lens' SqlParameter (Maybe Field)

The value of the parameter.

sqlParameter_name :: Lens' SqlParameter (Maybe Text)

The name of the parameter.

sqlParameter_typeHint :: Lens' SqlParameter (Maybe TypeHint)

A hint that specifies the correct object type for data type mapping. Possible values are as follows:

DATE - The corresponding String parameter value is sent as an object of DATE type to the database. The accepted format is YYYY-MM-DD.

DECIMAL - The corresponding String parameter value is sent as an object of DECIMAL type to the database.

JSON - The corresponding String parameter value is sent as an object of JSON type to the database.

TIME - The corresponding String parameter value is sent as an object of TIME type to the database. The accepted format is HH:MM:SS[.FFF].

TIMESTAMP - The corresponding String parameter value is sent as an object of TIMESTAMP type to the database. The accepted format is YYYY-MM-DD HH:MM:SS[.FFF].

UUID - The corresponding String parameter value is sent as an object of UUID type to the database.

UpdateResult

updateResult_generatedFields :: Lens' UpdateResult (Maybe [Field])

Values for fields generated during the request.