Amazonka.RDSData.RollbackTransaction

Description

Performs a rollback of a transaction. Rolling back a transaction cancels its changes.

Creating a Request

data RollbackTransaction

The request parameters represent the input of a request to perform a rollback of a transaction.

See: newRollbackTransaction smart constructor.

Constructors

RollbackTransaction'

Fields

resourceArn :: Text

The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

secretArn :: Text

The name or ARN of the secret that enables access to the DB cluster.

transactionId :: Text

The identifier of the transaction to roll back.

newRollbackTransaction

Arguments

:: Text
$sel:resourceArn:RollbackTransaction'

-> Text
$sel:secretArn:RollbackTransaction'

-> Text
$sel:transactionId:RollbackTransaction'

-> RollbackTransaction

Create a value of RollbackTransaction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:RollbackTransaction', rollbackTransaction_resourceArn - The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

$sel:secretArn:RollbackTransaction', rollbackTransaction_secretArn - The name or ARN of the secret that enables access to the DB cluster.

$sel:transactionId:RollbackTransaction', rollbackTransaction_transactionId - The identifier of the transaction to roll back.

Request Lenses

rollbackTransaction_resourceArn :: Lens' RollbackTransaction Text

The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

rollbackTransaction_secretArn :: Lens' RollbackTransaction Text

The name or ARN of the secret that enables access to the DB cluster.

rollbackTransaction_transactionId :: Lens' RollbackTransaction Text

The identifier of the transaction to roll back.

Destructuring the Response

data RollbackTransactionResponse

The response elements represent the output of a request to perform a rollback of a transaction.

See: newRollbackTransactionResponse smart constructor.

Constructors

RollbackTransactionResponse'

Fields

transactionStatus :: Maybe Text

The status of the rollback operation.

httpStatus :: Int

The response's http status code.

newRollbackTransactionResponse

Arguments

:: Int
$sel:httpStatus:RollbackTransactionResponse'

-> RollbackTransactionResponse

Create a value of RollbackTransactionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transactionStatus:RollbackTransactionResponse', rollbackTransactionResponse_transactionStatus - The status of the rollback operation.

$sel:httpStatus:RollbackTransactionResponse', rollbackTransactionResponse_httpStatus - The response's http status code.

Response Lenses

rollbackTransactionResponse_transactionStatus :: Lens' RollbackTransactionResponse (Maybe Text)

The status of the rollback operation.

rollbackTransactionResponse_httpStatus :: Lens' RollbackTransactionResponse Int

The response's http status code.