Amazonka.RDSData.Types.ColumnMetadata

Description

Documentation

data ColumnMetadata

Contains the metadata for a column.

See: newColumnMetadata smart constructor.

Constructors

ColumnMetadata'

Fields

typeName :: Maybe Text

The database-specific data type of the column.

isCaseSensitive :: Maybe Bool

A value that indicates whether the column is case-sensitive.

isCurrency :: Maybe Bool

A value that indicates whether the column contains currency values.

scale :: Maybe Int

The scale value of a decimal number column.

precision :: Maybe Int

The precision value of a decimal number column.

schemaName :: Maybe Text

The name of the schema that owns the table that includes the column.

isAutoIncrement :: Maybe Bool

A value that indicates whether the column increments automatically.

name :: Maybe Text

The name of the column.

arrayBaseColumnType :: Maybe Int

The type of the column.

type' :: Maybe Int

The type of the column.

isSigned :: Maybe Bool

A value that indicates whether an integer column is signed.

label :: Maybe Text

The label for the column.

nullable :: Maybe Int

A value that indicates whether the column is nullable.

tableName :: Maybe Text

The name of the table that includes the column.

newColumnMetadata :: ColumnMetadata

Create a value of ColumnMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:ColumnMetadata', columnMetadata_typeName - The database-specific data type of the column.

$sel:isCaseSensitive:ColumnMetadata', columnMetadata_isCaseSensitive - A value that indicates whether the column is case-sensitive.

$sel:isCurrency:ColumnMetadata', columnMetadata_isCurrency - A value that indicates whether the column contains currency values.

$sel:scale:ColumnMetadata', columnMetadata_scale - The scale value of a decimal number column.

$sel:precision:ColumnMetadata', columnMetadata_precision - The precision value of a decimal number column.

$sel:schemaName:ColumnMetadata', columnMetadata_schemaName - The name of the schema that owns the table that includes the column.

$sel:isAutoIncrement:ColumnMetadata', columnMetadata_isAutoIncrement - A value that indicates whether the column increments automatically.

$sel:name:ColumnMetadata', columnMetadata_name - The name of the column.

$sel:arrayBaseColumnType:ColumnMetadata', columnMetadata_arrayBaseColumnType - The type of the column.

$sel:type':ColumnMetadata', columnMetadata_type - The type of the column.

$sel:isSigned:ColumnMetadata', columnMetadata_isSigned - A value that indicates whether an integer column is signed.

$sel:label:ColumnMetadata', columnMetadata_label - The label for the column.

$sel:nullable:ColumnMetadata', columnMetadata_nullable - A value that indicates whether the column is nullable.

$sel:tableName:ColumnMetadata', columnMetadata_tableName - The name of the table that includes the column.

columnMetadata_typeName :: Lens' ColumnMetadata (Maybe Text)

The database-specific data type of the column.

columnMetadata_isCaseSensitive :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether the column is case-sensitive.

columnMetadata_isCurrency :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether the column contains currency values.

columnMetadata_scale :: Lens' ColumnMetadata (Maybe Int)

The scale value of a decimal number column.

columnMetadata_precision :: Lens' ColumnMetadata (Maybe Int)

The precision value of a decimal number column.

columnMetadata_schemaName :: Lens' ColumnMetadata (Maybe Text)

The name of the schema that owns the table that includes the column.

columnMetadata_isAutoIncrement :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether the column increments automatically.

columnMetadata_name :: Lens' ColumnMetadata (Maybe Text)

The name of the column.

columnMetadata_arrayBaseColumnType :: Lens' ColumnMetadata (Maybe Int)

The type of the column.

columnMetadata_type :: Lens' ColumnMetadata (Maybe Int)

The type of the column.

columnMetadata_isSigned :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether an integer column is signed.

columnMetadata_label :: Lens' ColumnMetadata (Maybe Text)

The label for the column.

columnMetadata_nullable :: Lens' ColumnMetadata (Maybe Int)

A value that indicates whether the column is nullable.

columnMetadata_tableName :: Lens' ColumnMetadata (Maybe Text)

The name of the table that includes the column.