Amazonka.RDSData.Types.Field

Description

Documentation

data Field

Contains a value.

See: newField smart constructor.

Constructors

Field'

Fields

doubleValue :: Maybe Double

A value of double data type.

stringValue :: Maybe Text

A value of string data type.

longValue :: Maybe Integer

A value of long data type.

booleanValue :: Maybe Bool

A value of Boolean data type.

arrayValue :: Maybe ArrayValue

An array of values.

blobValue :: Maybe Base64

A value of BLOB data type.

isNull :: Maybe Bool

A NULL value.

newField :: Field

Create a value of Field with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:doubleValue:Field', field_doubleValue - A value of double data type.

$sel:stringValue:Field', field_stringValue - A value of string data type.

$sel:longValue:Field', field_longValue - A value of long data type.

$sel:booleanValue:Field', field_booleanValue - A value of Boolean data type.

$sel:arrayValue:Field', field_arrayValue - An array of values.

$sel:blobValue:Field', field_blobValue - A value of BLOB data type.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:isNull:Field', field_isNull - A NULL value.

field_doubleValue :: Lens' Field (Maybe Double)

A value of double data type.

field_stringValue :: Lens' Field (Maybe Text)

A value of string data type.

field_longValue :: Lens' Field (Maybe Integer)

A value of long data type.

field_booleanValue :: Lens' Field (Maybe Bool)

A value of Boolean data type.

field_arrayValue :: Lens' Field (Maybe ArrayValue)

An array of values.

field_blobValue :: Lens' Field (Maybe ByteString)

A value of BLOB data type.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

field_isNull :: Lens' Field (Maybe Bool)

A NULL value.