Amazonka.Redshift.BatchModifyClusterSnapshots

Description

Modifies the settings for a set of cluster snapshots.

Creating a Request

data BatchModifyClusterSnapshots

See: newBatchModifyClusterSnapshots smart constructor.

Constructors

BatchModifyClusterSnapshots'

Fields

manualSnapshotRetentionPeriod :: Maybe Int

The number of days that a manual snapshot is retained. If you specify the value -1, the manual snapshot is retained indefinitely.

The number must be either -1 or an integer between 1 and 3,653.

If you decrease the manual snapshot retention period from its current value, existing manual snapshots that fall outside of the new retention period will return an error. If you want to suppress the errors and delete the snapshots, use the force option.

force :: Maybe Bool

A boolean value indicating whether to override an exception if the retention period has passed.

snapshotIdentifierList :: [Text]

A list of snapshot identifiers you want to modify.

newBatchModifyClusterSnapshots :: BatchModifyClusterSnapshots

Create a value of BatchModifyClusterSnapshots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manualSnapshotRetentionPeriod:BatchModifyClusterSnapshots', batchModifyClusterSnapshots_manualSnapshotRetentionPeriod - The number of days that a manual snapshot is retained. If you specify the value -1, the manual snapshot is retained indefinitely.

The number must be either -1 or an integer between 1 and 3,653.

If you decrease the manual snapshot retention period from its current value, existing manual snapshots that fall outside of the new retention period will return an error. If you want to suppress the errors and delete the snapshots, use the force option.

$sel:force:BatchModifyClusterSnapshots', batchModifyClusterSnapshots_force - A boolean value indicating whether to override an exception if the retention period has passed.

$sel:snapshotIdentifierList:BatchModifyClusterSnapshots', batchModifyClusterSnapshots_snapshotIdentifierList - A list of snapshot identifiers you want to modify.

Request Lenses

batchModifyClusterSnapshots_manualSnapshotRetentionPeriod :: Lens' BatchModifyClusterSnapshots (Maybe Int)

The number of days that a manual snapshot is retained. If you specify the value -1, the manual snapshot is retained indefinitely.

The number must be either -1 or an integer between 1 and 3,653.

If you decrease the manual snapshot retention period from its current value, existing manual snapshots that fall outside of the new retention period will return an error. If you want to suppress the errors and delete the snapshots, use the force option.

batchModifyClusterSnapshots_force :: Lens' BatchModifyClusterSnapshots (Maybe Bool)

A boolean value indicating whether to override an exception if the retention period has passed.

batchModifyClusterSnapshots_snapshotIdentifierList :: Lens' BatchModifyClusterSnapshots [Text]

A list of snapshot identifiers you want to modify.

Destructuring the Response

data BatchModifyClusterSnapshotsResponse

See: newBatchModifyClusterSnapshotsResponse smart constructor.

Constructors

BatchModifyClusterSnapshotsResponse'

Fields

resources :: Maybe [Text]

A list of the snapshots that were modified.

errors :: Maybe [SnapshotErrorMessage]

A list of any errors returned.

httpStatus :: Int

The response's http status code.

newBatchModifyClusterSnapshotsResponse

Arguments

:: Int
$sel:httpStatus:BatchModifyClusterSnapshotsResponse'

-> BatchModifyClusterSnapshotsResponse

Create a value of BatchModifyClusterSnapshotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resources:BatchModifyClusterSnapshotsResponse', batchModifyClusterSnapshotsResponse_resources - A list of the snapshots that were modified.

$sel:errors:BatchModifyClusterSnapshotsResponse', batchModifyClusterSnapshotsResponse_errors - A list of any errors returned.

$sel:httpStatus:BatchModifyClusterSnapshotsResponse', batchModifyClusterSnapshotsResponse_httpStatus - The response's http status code.

Response Lenses

batchModifyClusterSnapshotsResponse_resources :: Lens' BatchModifyClusterSnapshotsResponse (Maybe [Text])

A list of the snapshots that were modified.

batchModifyClusterSnapshotsResponse_errors :: Lens' BatchModifyClusterSnapshotsResponse (Maybe [SnapshotErrorMessage])

A list of any errors returned.

batchModifyClusterSnapshotsResponse_httpStatus :: Lens' BatchModifyClusterSnapshotsResponse Int

The response's http status code.