Amazonka.Redshift.CreateClusterSnapshot

Description

Creates a manual snapshot of the specified cluster. The cluster must be in the available state.

For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Cluster Management Guide.

Creating a Request

data CreateClusterSnapshot

See: newCreateClusterSnapshot smart constructor.

Constructors

CreateClusterSnapshot'

Fields

manualSnapshotRetentionPeriod :: Maybe Int

The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.

tags :: Maybe [Tag]

A list of tag instances.

snapshotIdentifier :: Text

A unique identifier for the snapshot that you are requesting. This identifier must be unique for all snapshots within the Amazon Web Services account.

Constraints:

Cannot be null, empty, or blank

Must contain from 1 to 255 alphanumeric characters or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

clusterIdentifier :: Text

The cluster identifier for which you want a snapshot.

newCreateClusterSnapshot

Arguments

:: Text
$sel:snapshotIdentifier:CreateClusterSnapshot'

-> Text
$sel:clusterIdentifier:CreateClusterSnapshot'

-> CreateClusterSnapshot

Create a value of CreateClusterSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manualSnapshotRetentionPeriod:CreateClusterSnapshot', createClusterSnapshot_manualSnapshotRetentionPeriod - The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.

$sel:tags:CreateClusterSnapshot', createClusterSnapshot_tags - A list of tag instances.

$sel:snapshotIdentifier:CreateClusterSnapshot', createClusterSnapshot_snapshotIdentifier - A unique identifier for the snapshot that you are requesting. This identifier must be unique for all snapshots within the Amazon Web Services account.

Constraints:

Cannot be null, empty, or blank

Must contain from 1 to 255 alphanumeric characters or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

$sel:clusterIdentifier:CreateClusterSnapshot', createClusterSnapshot_clusterIdentifier - The cluster identifier for which you want a snapshot.

Request Lenses

createClusterSnapshot_manualSnapshotRetentionPeriod :: Lens' CreateClusterSnapshot (Maybe Int)

The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.

createClusterSnapshot_tags :: Lens' CreateClusterSnapshot (Maybe [Tag])

A list of tag instances.

createClusterSnapshot_snapshotIdentifier :: Lens' CreateClusterSnapshot Text

A unique identifier for the snapshot that you are requesting. This identifier must be unique for all snapshots within the Amazon Web Services account.

Constraints:

Cannot be null, empty, or blank

Must contain from 1 to 255 alphanumeric characters or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

createClusterSnapshot_clusterIdentifier :: Lens' CreateClusterSnapshot Text

The cluster identifier for which you want a snapshot.

Destructuring the Response

data CreateClusterSnapshotResponse

See: newCreateClusterSnapshotResponse smart constructor.

Constructors

CreateClusterSnapshotResponse'

Fields

snapshot :: Maybe Snapshot

httpStatus :: Int

The response's http status code.

newCreateClusterSnapshotResponse

Arguments

:: Int
$sel:httpStatus:CreateClusterSnapshotResponse'

-> CreateClusterSnapshotResponse

Create a value of CreateClusterSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshot:CreateClusterSnapshotResponse', createClusterSnapshotResponse_snapshot - Undocumented member.

$sel:httpStatus:CreateClusterSnapshotResponse', createClusterSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

createClusterSnapshotResponse_snapshot :: Lens' CreateClusterSnapshotResponse (Maybe Snapshot)

Undocumented member.

createClusterSnapshotResponse_httpStatus :: Lens' CreateClusterSnapshotResponse Int

The response's http status code.