Amazonka.Redshift.CreateSnapshotCopyGrant

Description

Creates a snapshot copy grant that permits Amazon Redshift to use a customer master key (CMK) from Key Management Service (KMS) to encrypt copied snapshots in a destination region.

For more information about managing snapshot copy grants, go to Amazon Redshift Database Encryption in the Amazon Redshift Cluster Management Guide.

Creating a Request

data CreateSnapshotCopyGrant

The result of the CreateSnapshotCopyGrant action.

See: newCreateSnapshotCopyGrant smart constructor.

Constructors

CreateSnapshotCopyGrant'

Fields

kmsKeyId :: Maybe Text

The unique identifier of the customer master key (CMK) to which to grant Amazon Redshift permission. If no key is specified, the default key is used.

tags :: Maybe [Tag]

A list of tag instances.

snapshotCopyGrantName :: Text

The name of the snapshot copy grant. This name must be unique in the region for the Amazon Web Services account.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

newCreateSnapshotCopyGrant

Arguments

:: Text
$sel:snapshotCopyGrantName:CreateSnapshotCopyGrant'

-> CreateSnapshotCopyGrant

Create a value of CreateSnapshotCopyGrant with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:CreateSnapshotCopyGrant', createSnapshotCopyGrant_kmsKeyId - The unique identifier of the customer master key (CMK) to which to grant Amazon Redshift permission. If no key is specified, the default key is used.

$sel:tags:CreateSnapshotCopyGrant', createSnapshotCopyGrant_tags - A list of tag instances.

$sel:snapshotCopyGrantName:CreateSnapshotCopyGrant', createSnapshotCopyGrant_snapshotCopyGrantName - The name of the snapshot copy grant. This name must be unique in the region for the Amazon Web Services account.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

Request Lenses

createSnapshotCopyGrant_kmsKeyId :: Lens' CreateSnapshotCopyGrant (Maybe Text)

The unique identifier of the customer master key (CMK) to which to grant Amazon Redshift permission. If no key is specified, the default key is used.

createSnapshotCopyGrant_tags :: Lens' CreateSnapshotCopyGrant (Maybe [Tag])

A list of tag instances.

createSnapshotCopyGrant_snapshotCopyGrantName :: Lens' CreateSnapshotCopyGrant Text

The name of the snapshot copy grant. This name must be unique in the region for the Amazon Web Services account.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

Destructuring the Response

data CreateSnapshotCopyGrantResponse

See: newCreateSnapshotCopyGrantResponse smart constructor.

Constructors

CreateSnapshotCopyGrantResponse'

Fields

snapshotCopyGrant :: Maybe SnapshotCopyGrant

httpStatus :: Int

The response's http status code.

newCreateSnapshotCopyGrantResponse

Arguments

:: Int
$sel:httpStatus:CreateSnapshotCopyGrantResponse'

-> CreateSnapshotCopyGrantResponse

Create a value of CreateSnapshotCopyGrantResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotCopyGrant:CreateSnapshotCopyGrantResponse', createSnapshotCopyGrantResponse_snapshotCopyGrant - Undocumented member.

$sel:httpStatus:CreateSnapshotCopyGrantResponse', createSnapshotCopyGrantResponse_httpStatus - The response's http status code.

Response Lenses

createSnapshotCopyGrantResponse_snapshotCopyGrant :: Lens' CreateSnapshotCopyGrantResponse (Maybe SnapshotCopyGrant)

Undocumented member.

createSnapshotCopyGrantResponse_httpStatus :: Lens' CreateSnapshotCopyGrantResponse Int

The response's http status code.