Amazonka.Redshift.DeauthorizeDataShare

Description

From the producer account, removes authorization from the specified datashare.

Creating a Request

data DeauthorizeDataShare

See: newDeauthorizeDataShare smart constructor.

Constructors

DeauthorizeDataShare'

Fields

dataShareArn :: Text

The Amazon Resource Name (ARN) of the datashare to remove authorization from.

consumerIdentifier :: Text

The identifier of the data consumer that is to have authorization removed from the datashare. This identifier is an AWS account ID.

newDeauthorizeDataShare

Arguments

:: Text
$sel:dataShareArn:DeauthorizeDataShare'

-> Text
$sel:consumerIdentifier:DeauthorizeDataShare'

-> DeauthorizeDataShare

Create a value of DeauthorizeDataShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataShareArn:DeauthorizeDataShare', deauthorizeDataShare_dataShareArn - The Amazon Resource Name (ARN) of the datashare to remove authorization from.

$sel:consumerIdentifier:DeauthorizeDataShare', deauthorizeDataShare_consumerIdentifier - The identifier of the data consumer that is to have authorization removed from the datashare. This identifier is an AWS account ID.

Request Lenses

deauthorizeDataShare_dataShareArn :: Lens' DeauthorizeDataShare Text

The Amazon Resource Name (ARN) of the datashare to remove authorization from.

deauthorizeDataShare_consumerIdentifier :: Lens' DeauthorizeDataShare Text

The identifier of the data consumer that is to have authorization removed from the datashare. This identifier is an AWS account ID.

Destructuring the Response

data DataShare

See: newDataShare smart constructor.

Constructors

DataShare'

Fields

producerArn :: Maybe Text

The Amazon Resource Name (ARN) of the producer.

dataShareAssociations :: Maybe [DataShareAssociation]

A value that specifies when the datashare has an association between a producer and data consumers.

dataShareArn :: Maybe Text

An Amazon Resource Name (ARN) that references the datashare that is owned by a specific namespace of the producer cluster. A datashare ARN is in the arn:aws:redshift:{region}:{account-id}:{datashare}:{namespace-guid}/{datashare-name} format.

allowPubliclyAccessibleConsumers :: Maybe Bool

A value that specifies whether the datashare can be shared to a publicly accessible cluster.

newDataShare :: DataShare

Create a value of DataShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:producerArn:DataShare', dataShare_producerArn - The Amazon Resource Name (ARN) of the producer.

$sel:dataShareAssociations:DataShare', dataShare_dataShareAssociations - A value that specifies when the datashare has an association between a producer and data consumers.

$sel:dataShareArn:DataShare', dataShare_dataShareArn - An Amazon Resource Name (ARN) that references the datashare that is owned by a specific namespace of the producer cluster. A datashare ARN is in the arn:aws:redshift:{region}:{account-id}:{datashare}:{namespace-guid}/{datashare-name} format.

$sel:allowPubliclyAccessibleConsumers:DataShare', dataShare_allowPubliclyAccessibleConsumers - A value that specifies whether the datashare can be shared to a publicly accessible cluster.

Response Lenses

dataShare_producerArn :: Lens' DataShare (Maybe Text)

The Amazon Resource Name (ARN) of the producer.

dataShare_dataShareAssociations :: Lens' DataShare (Maybe [DataShareAssociation])

A value that specifies when the datashare has an association between a producer and data consumers.

dataShare_dataShareArn :: Lens' DataShare (Maybe Text)

An Amazon Resource Name (ARN) that references the datashare that is owned by a specific namespace of the producer cluster. A datashare ARN is in the arn:aws:redshift:{region}:{account-id}:{datashare}:{namespace-guid}/{datashare-name} format.

dataShare_allowPubliclyAccessibleConsumers :: Lens' DataShare (Maybe Bool)

A value that specifies whether the datashare can be shared to a publicly accessible cluster.