Amazonka.Redshift.DescribeClusterSnapshots

Description

Returns one or more snapshot objects, which contain metadata about your cluster snapshots. By default, this operation returns information about all snapshots of all clusters that are owned by your Amazon Web Services account. No information is returned for snapshots owned by inactive Amazon Web Services accounts.

If you specify both tag keys and tag values in the same request, Amazon Redshift returns all snapshots that match any combination of the specified keys and values. For example, if you have owner and environment for tag keys, and admin and test for tag values, all snapshots that have any combination of those values are returned. Only snapshots that you own are returned in the response; shared snapshots are not returned with the tag key and tag value request parameters.

If both tag keys and values are omitted from the request, snapshots are returned regardless of whether they have tag keys or values associated with them.

This operation returns paginated results.

Creating a Request

data DescribeClusterSnapshots

See: newDescribeClusterSnapshots smart constructor.

Constructors

DescribeClusterSnapshots'

Fields

snapshotIdentifier :: Maybe Text

The snapshot identifier of the snapshot about which to return information.

tagValues :: Maybe [Text]

A tag value or values for which you want to return all matching cluster snapshots that are associated with the specified tag value or values. For example, suppose that you have snapshots that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the snapshots that have either or both of these tag values associated with them.

clusterExists :: Maybe Bool

A value that indicates whether to return snapshots only for an existing cluster. You can perform table-level restore only by using a snapshot of an existing cluster, that is, a cluster that has not been deleted. Values for this parameter work as follows:

If ClusterExists is set to true, ClusterIdentifier is required.

If ClusterExists is set to false and ClusterIdentifier isn't specified, all snapshots associated with deleted clusters (orphaned snapshots) are returned.

If ClusterExists is set to false and ClusterIdentifier is specified for a deleted cluster, snapshots associated with that cluster are returned.

If ClusterExists is set to false and ClusterIdentifier is specified for an existing cluster, no snapshots are returned.

startTime :: Maybe ISO8601

A value that requests only snapshots created at or after the specified time. The time value is specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page.

Example: 2012-07-16T18:00:00Z

tagKeys :: Maybe [Text]

A tag key or keys for which you want to return all matching cluster snapshots that are associated with the specified key or keys. For example, suppose that you have snapshots that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the snapshots that have either or both of these tag keys associated with them.

clusterIdentifier :: Maybe Text

The identifier of the cluster which generated the requested snapshots.

snapshotType :: Maybe Text

The type of snapshots for which you are requesting information. By default, snapshots of all types are returned.

Valid Values: automated | manual

sortingEntities :: Maybe [SnapshotSortingEntity]

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterSnapshots request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

endTime :: Maybe ISO8601

A time value that requests only snapshots created at or before the specified time. The time value is specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page.

Example: 2012-07-16T18:00:00Z

ownerAccount :: Maybe Text

The Amazon Web Services account used to create or copy the snapshot. Use this field to filter the results to snapshots owned by a particular account. To describe snapshots you own, either specify your Amazon Web Services account, or do not specify the parameter.

newDescribeClusterSnapshots :: DescribeClusterSnapshots

Create a value of DescribeClusterSnapshots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotIdentifier:DescribeClusterSnapshots', describeClusterSnapshots_snapshotIdentifier - The snapshot identifier of the snapshot about which to return information.

$sel:tagValues:DescribeClusterSnapshots', describeClusterSnapshots_tagValues - A tag value or values for which you want to return all matching cluster snapshots that are associated with the specified tag value or values. For example, suppose that you have snapshots that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the snapshots that have either or both of these tag values associated with them.

$sel:clusterExists:DescribeClusterSnapshots', describeClusterSnapshots_clusterExists - A value that indicates whether to return snapshots only for an existing cluster. You can perform table-level restore only by using a snapshot of an existing cluster, that is, a cluster that has not been deleted. Values for this parameter work as follows:

If ClusterExists is set to true, ClusterIdentifier is required.

If ClusterExists is set to false and ClusterIdentifier isn't specified, all snapshots associated with deleted clusters (orphaned snapshots) are returned.

If ClusterExists is set to false and ClusterIdentifier is specified for a deleted cluster, snapshots associated with that cluster are returned.

If ClusterExists is set to false and ClusterIdentifier is specified for an existing cluster, no snapshots are returned.

$sel:startTime:DescribeClusterSnapshots', describeClusterSnapshots_startTime - A value that requests only snapshots created at or after the specified time. The time value is specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page.

Example: 2012-07-16T18:00:00Z

$sel:tagKeys:DescribeClusterSnapshots', describeClusterSnapshots_tagKeys - A tag key or keys for which you want to return all matching cluster snapshots that are associated with the specified key or keys. For example, suppose that you have snapshots that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the snapshots that have either or both of these tag keys associated with them.

$sel:clusterIdentifier:DescribeClusterSnapshots', describeClusterSnapshots_clusterIdentifier - The identifier of the cluster which generated the requested snapshots.

$sel:snapshotType:DescribeClusterSnapshots', describeClusterSnapshots_snapshotType - The type of snapshots for which you are requesting information. By default, snapshots of all types are returned.

Valid Values: automated | manual

$sel:sortingEntities:DescribeClusterSnapshots', describeClusterSnapshots_sortingEntities -

$sel:marker:DescribeClusterSnapshots', describeClusterSnapshots_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterSnapshots request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:maxRecords:DescribeClusterSnapshots', describeClusterSnapshots_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

$sel:endTime:DescribeClusterSnapshots', describeClusterSnapshots_endTime - A time value that requests only snapshots created at or before the specified time. The time value is specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page.

Example: 2012-07-16T18:00:00Z

$sel:ownerAccount:DescribeClusterSnapshots', describeClusterSnapshots_ownerAccount - The Amazon Web Services account used to create or copy the snapshot. Use this field to filter the results to snapshots owned by a particular account. To describe snapshots you own, either specify your Amazon Web Services account, or do not specify the parameter.

Request Lenses

describeClusterSnapshots_snapshotIdentifier :: Lens' DescribeClusterSnapshots (Maybe Text)

The snapshot identifier of the snapshot about which to return information.

describeClusterSnapshots_tagValues :: Lens' DescribeClusterSnapshots (Maybe [Text])

A tag value or values for which you want to return all matching cluster snapshots that are associated with the specified tag value or values. For example, suppose that you have snapshots that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the snapshots that have either or both of these tag values associated with them.

describeClusterSnapshots_clusterExists :: Lens' DescribeClusterSnapshots (Maybe Bool)

A value that indicates whether to return snapshots only for an existing cluster. You can perform table-level restore only by using a snapshot of an existing cluster, that is, a cluster that has not been deleted. Values for this parameter work as follows:

If ClusterExists is set to true, ClusterIdentifier is required.

If ClusterExists is set to false and ClusterIdentifier isn't specified, all snapshots associated with deleted clusters (orphaned snapshots) are returned.

If ClusterExists is set to false and ClusterIdentifier is specified for a deleted cluster, snapshots associated with that cluster are returned.

If ClusterExists is set to false and ClusterIdentifier is specified for an existing cluster, no snapshots are returned.

describeClusterSnapshots_startTime :: Lens' DescribeClusterSnapshots (Maybe UTCTime)

A value that requests only snapshots created at or after the specified time. The time value is specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page.

Example: 2012-07-16T18:00:00Z

describeClusterSnapshots_tagKeys :: Lens' DescribeClusterSnapshots (Maybe [Text])

A tag key or keys for which you want to return all matching cluster snapshots that are associated with the specified key or keys. For example, suppose that you have snapshots that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the snapshots that have either or both of these tag keys associated with them.

describeClusterSnapshots_clusterIdentifier :: Lens' DescribeClusterSnapshots (Maybe Text)

The identifier of the cluster which generated the requested snapshots.

describeClusterSnapshots_snapshotType :: Lens' DescribeClusterSnapshots (Maybe Text)

The type of snapshots for which you are requesting information. By default, snapshots of all types are returned.

Valid Values: automated | manual

describeClusterSnapshots_sortingEntities :: Lens' DescribeClusterSnapshots (Maybe [SnapshotSortingEntity])

describeClusterSnapshots_marker :: Lens' DescribeClusterSnapshots (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterSnapshots request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeClusterSnapshots_maxRecords :: Lens' DescribeClusterSnapshots (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

describeClusterSnapshots_endTime :: Lens' DescribeClusterSnapshots (Maybe UTCTime)

A time value that requests only snapshots created at or before the specified time. The time value is specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page.

Example: 2012-07-16T18:00:00Z

describeClusterSnapshots_ownerAccount :: Lens' DescribeClusterSnapshots (Maybe Text)

The Amazon Web Services account used to create or copy the snapshot. Use this field to filter the results to snapshots owned by a particular account. To describe snapshots you own, either specify your Amazon Web Services account, or do not specify the parameter.

Destructuring the Response

data DescribeClusterSnapshotsResponse

Contains the output from the DescribeClusterSnapshots action.

See: newDescribeClusterSnapshotsResponse smart constructor.

Constructors

DescribeClusterSnapshotsResponse'

Fields

snapshots :: Maybe [Snapshot]

A list of Snapshot instances.

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

httpStatus :: Int

The response's http status code.

newDescribeClusterSnapshotsResponse

Arguments

:: Int
$sel:httpStatus:DescribeClusterSnapshotsResponse'

-> DescribeClusterSnapshotsResponse

Create a value of DescribeClusterSnapshotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshots:DescribeClusterSnapshotsResponse', describeClusterSnapshotsResponse_snapshots - A list of Snapshot instances.

$sel:marker:DescribeClusterSnapshots', describeClusterSnapshotsResponse_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

$sel:httpStatus:DescribeClusterSnapshotsResponse', describeClusterSnapshotsResponse_httpStatus - The response's http status code.

Response Lenses

describeClusterSnapshotsResponse_snapshots :: Lens' DescribeClusterSnapshotsResponse (Maybe [Snapshot])

A list of Snapshot instances.

describeClusterSnapshotsResponse_marker :: Lens' DescribeClusterSnapshotsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

describeClusterSnapshotsResponse_httpStatus :: Lens' DescribeClusterSnapshotsResponse Int

The response's http status code.