Amazonka.Redshift.DescribeClusterTracks

Description

Returns a list of all the available maintenance tracks.

This operation returns paginated results.

Creating a Request

data DescribeClusterTracks

See: newDescribeClusterTracks smart constructor.

Constructors

DescribeClusterTracks'

Fields

maintenanceTrackName :: Maybe Text

The name of the maintenance track.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterTracks request exceed the value specified in MaxRecords, Amazon Redshift returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

maxRecords :: Maybe Int

An integer value for the maximum number of maintenance tracks to return.

newDescribeClusterTracks :: DescribeClusterTracks

Create a value of DescribeClusterTracks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maintenanceTrackName:DescribeClusterTracks', describeClusterTracks_maintenanceTrackName - The name of the maintenance track.

$sel:marker:DescribeClusterTracks', describeClusterTracks_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterTracks request exceed the value specified in MaxRecords, Amazon Redshift returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:maxRecords:DescribeClusterTracks', describeClusterTracks_maxRecords - An integer value for the maximum number of maintenance tracks to return.

Request Lenses

describeClusterTracks_maintenanceTrackName :: Lens' DescribeClusterTracks (Maybe Text)

The name of the maintenance track.

describeClusterTracks_marker :: Lens' DescribeClusterTracks (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterTracks request exceed the value specified in MaxRecords, Amazon Redshift returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeClusterTracks_maxRecords :: Lens' DescribeClusterTracks (Maybe Int)

An integer value for the maximum number of maintenance tracks to return.

Destructuring the Response

data DescribeClusterTracksResponse

See: newDescribeClusterTracksResponse smart constructor.

Constructors

DescribeClusterTracksResponse'

Fields

maintenanceTracks :: Maybe [MaintenanceTrack]

A list of maintenance tracks output by the DescribeClusterTracks operation.

marker :: Maybe Text

The starting point to return a set of response tracklist records. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

httpStatus :: Int

The response's http status code.

newDescribeClusterTracksResponse

Arguments

:: Int
$sel:httpStatus:DescribeClusterTracksResponse'

-> DescribeClusterTracksResponse

Create a value of DescribeClusterTracksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maintenanceTracks:DescribeClusterTracksResponse', describeClusterTracksResponse_maintenanceTracks - A list of maintenance tracks output by the DescribeClusterTracks operation.

$sel:marker:DescribeClusterTracks', describeClusterTracksResponse_marker - The starting point to return a set of response tracklist records. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:httpStatus:DescribeClusterTracksResponse', describeClusterTracksResponse_httpStatus - The response's http status code.

Response Lenses

describeClusterTracksResponse_maintenanceTracks :: Lens' DescribeClusterTracksResponse (Maybe [MaintenanceTrack])

A list of maintenance tracks output by the DescribeClusterTracks operation.

describeClusterTracksResponse_marker :: Lens' DescribeClusterTracksResponse (Maybe Text)

The starting point to return a set of response tracklist records. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeClusterTracksResponse_httpStatus :: Lens' DescribeClusterTracksResponse Int

The response's http status code.