Amazonka.Redshift.DescribeDataSharesForConsumer

Description

Returns a list of datashares where the account identifier being called is a consumer account identifier.

Creating a Request

data DescribeDataSharesForConsumer

See: newDescribeDataSharesForConsumer smart constructor.

Constructors

DescribeDataSharesForConsumer'

Fields

status :: Maybe DataShareStatusForConsumer

An identifier giving the status of a datashare in the consumer cluster. If this field is specified, Amazon Redshift returns the list of datashares that have the specified status.

consumerArn :: Maybe Text

The Amazon Resource Name (ARN) of the consumer that returns in the list of datashares.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataSharesForConsumer request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

newDescribeDataSharesForConsumer :: DescribeDataSharesForConsumer

Create a value of DescribeDataSharesForConsumer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeDataSharesForConsumer', describeDataSharesForConsumer_status - An identifier giving the status of a datashare in the consumer cluster. If this field is specified, Amazon Redshift returns the list of datashares that have the specified status.

$sel:consumerArn:DescribeDataSharesForConsumer', describeDataSharesForConsumer_consumerArn - The Amazon Resource Name (ARN) of the consumer that returns in the list of datashares.

$sel:marker:DescribeDataSharesForConsumer', describeDataSharesForConsumer_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataSharesForConsumer request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:maxRecords:DescribeDataSharesForConsumer', describeDataSharesForConsumer_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Request Lenses

describeDataSharesForConsumer_status :: Lens' DescribeDataSharesForConsumer (Maybe DataShareStatusForConsumer)

An identifier giving the status of a datashare in the consumer cluster. If this field is specified, Amazon Redshift returns the list of datashares that have the specified status.

describeDataSharesForConsumer_consumerArn :: Lens' DescribeDataSharesForConsumer (Maybe Text)

The Amazon Resource Name (ARN) of the consumer that returns in the list of datashares.

describeDataSharesForConsumer_marker :: Lens' DescribeDataSharesForConsumer (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataSharesForConsumer request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeDataSharesForConsumer_maxRecords :: Lens' DescribeDataSharesForConsumer (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Destructuring the Response

data DescribeDataSharesForConsumerResponse

See: newDescribeDataSharesForConsumerResponse smart constructor.

Constructors

DescribeDataSharesForConsumerResponse'

Fields

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataSharesForConsumer request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

dataShares :: Maybe [DataShare]

Shows the results of datashares available for consumers.

httpStatus :: Int

The response's http status code.

newDescribeDataSharesForConsumerResponse

Arguments

:: Int
$sel:httpStatus:DescribeDataSharesForConsumerResponse'

-> DescribeDataSharesForConsumerResponse

Create a value of DescribeDataSharesForConsumerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeDataSharesForConsumer', describeDataSharesForConsumerResponse_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataSharesForConsumer request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:dataShares:DescribeDataSharesForConsumerResponse', describeDataSharesForConsumerResponse_dataShares - Shows the results of datashares available for consumers.

$sel:httpStatus:DescribeDataSharesForConsumerResponse', describeDataSharesForConsumerResponse_httpStatus - The response's http status code.

Response Lenses

describeDataSharesForConsumerResponse_marker :: Lens' DescribeDataSharesForConsumerResponse (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataSharesForConsumer request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeDataSharesForConsumerResponse_dataShares :: Lens' DescribeDataSharesForConsumerResponse (Maybe [DataShare])

Shows the results of datashares available for consumers.

describeDataSharesForConsumerResponse_httpStatus :: Lens' DescribeDataSharesForConsumerResponse Int

The response's http status code.