Amazonka.Redshift.DescribeHsmClientCertificates

Description

Returns information about the specified HSM client certificate. If no certificate ID is specified, returns information about all the HSM certificates owned by your Amazon Web Services account.

If you specify both tag keys and tag values in the same request, Amazon Redshift returns all HSM client certificates that match any combination of the specified keys and values. For example, if you have owner and environment for tag keys, and admin and test for tag values, all HSM client certificates that have any combination of those values are returned.

If both tag keys and values are omitted from the request, HSM client certificates are returned regardless of whether they have tag keys or values associated with them.

This operation returns paginated results.

Creating a Request

data DescribeHsmClientCertificates

See: newDescribeHsmClientCertificates smart constructor.

Constructors

DescribeHsmClientCertificates'

Fields

tagValues :: Maybe [Text]

A tag value or values for which you want to return all matching HSM client certificates that are associated with the specified tag value or values. For example, suppose that you have HSM client certificates that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the HSM client certificates that have either or both of these tag values associated with them.

tagKeys :: Maybe [Text]

A tag key or keys for which you want to return all matching HSM client certificates that are associated with the specified key or keys. For example, suppose that you have HSM client certificates that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the HSM client certificates that have either or both of these tag keys associated with them.

hsmClientCertificateIdentifier :: Maybe Text

The identifier of a specific HSM client certificate for which you want information. If no identifier is specified, information is returned for all HSM client certificates owned by your Amazon Web Services account.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeHsmClientCertificates request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

newDescribeHsmClientCertificates :: DescribeHsmClientCertificates

Create a value of DescribeHsmClientCertificates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagValues:DescribeHsmClientCertificates', describeHsmClientCertificates_tagValues - A tag value or values for which you want to return all matching HSM client certificates that are associated with the specified tag value or values. For example, suppose that you have HSM client certificates that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the HSM client certificates that have either or both of these tag values associated with them.

$sel:tagKeys:DescribeHsmClientCertificates', describeHsmClientCertificates_tagKeys - A tag key or keys for which you want to return all matching HSM client certificates that are associated with the specified key or keys. For example, suppose that you have HSM client certificates that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the HSM client certificates that have either or both of these tag keys associated with them.

$sel:hsmClientCertificateIdentifier:DescribeHsmClientCertificates', describeHsmClientCertificates_hsmClientCertificateIdentifier - The identifier of a specific HSM client certificate for which you want information. If no identifier is specified, information is returned for all HSM client certificates owned by your Amazon Web Services account.

$sel:marker:DescribeHsmClientCertificates', describeHsmClientCertificates_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeHsmClientCertificates request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:maxRecords:DescribeHsmClientCertificates', describeHsmClientCertificates_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

Request Lenses

describeHsmClientCertificates_tagValues :: Lens' DescribeHsmClientCertificates (Maybe [Text])

A tag value or values for which you want to return all matching HSM client certificates that are associated with the specified tag value or values. For example, suppose that you have HSM client certificates that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the HSM client certificates that have either or both of these tag values associated with them.

describeHsmClientCertificates_tagKeys :: Lens' DescribeHsmClientCertificates (Maybe [Text])

A tag key or keys for which you want to return all matching HSM client certificates that are associated with the specified key or keys. For example, suppose that you have HSM client certificates that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the HSM client certificates that have either or both of these tag keys associated with them.

describeHsmClientCertificates_hsmClientCertificateIdentifier :: Lens' DescribeHsmClientCertificates (Maybe Text)

The identifier of a specific HSM client certificate for which you want information. If no identifier is specified, information is returned for all HSM client certificates owned by your Amazon Web Services account.

describeHsmClientCertificates_marker :: Lens' DescribeHsmClientCertificates (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeHsmClientCertificates request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeHsmClientCertificates_maxRecords :: Lens' DescribeHsmClientCertificates (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

Destructuring the Response

data DescribeHsmClientCertificatesResponse

See: newDescribeHsmClientCertificatesResponse smart constructor.

Constructors

DescribeHsmClientCertificatesResponse'

Fields

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

hsmClientCertificates :: Maybe [HsmClientCertificate]

A list of the identifiers for one or more HSM client certificates used by Amazon Redshift clusters to store and retrieve database encryption keys in an HSM.

httpStatus :: Int

The response's http status code.

newDescribeHsmClientCertificatesResponse

Arguments

:: Int
$sel:httpStatus:DescribeHsmClientCertificatesResponse'

-> DescribeHsmClientCertificatesResponse

Create a value of DescribeHsmClientCertificatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeHsmClientCertificates', describeHsmClientCertificatesResponse_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

$sel:hsmClientCertificates:DescribeHsmClientCertificatesResponse', describeHsmClientCertificatesResponse_hsmClientCertificates - A list of the identifiers for one or more HSM client certificates used by Amazon Redshift clusters to store and retrieve database encryption keys in an HSM.

$sel:httpStatus:DescribeHsmClientCertificatesResponse', describeHsmClientCertificatesResponse_httpStatus - The response's http status code.

Response Lenses

describeHsmClientCertificatesResponse_marker :: Lens' DescribeHsmClientCertificatesResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

describeHsmClientCertificatesResponse_hsmClientCertificates :: Lens' DescribeHsmClientCertificatesResponse (Maybe [HsmClientCertificate])

A list of the identifiers for one or more HSM client certificates used by Amazon Redshift clusters to store and retrieve database encryption keys in an HSM.

describeHsmClientCertificatesResponse_httpStatus :: Lens' DescribeHsmClientCertificatesResponse Int

The response's http status code.