Amazonka.Redshift.DescribeLoggingStatus

Description

Describes whether information, such as queries and connection attempts, is being logged for the specified Amazon Redshift cluster.

Creating a Request

data DescribeLoggingStatus

See: newDescribeLoggingStatus smart constructor.

Constructors

DescribeLoggingStatus'

Fields

clusterIdentifier :: Text

The identifier of the cluster from which to get the logging status.

Example: examplecluster

newDescribeLoggingStatus

Arguments

:: Text
$sel:clusterIdentifier:DescribeLoggingStatus'

-> DescribeLoggingStatus

Create a value of DescribeLoggingStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterIdentifier:DescribeLoggingStatus', describeLoggingStatus_clusterIdentifier - The identifier of the cluster from which to get the logging status.

Example: examplecluster

Request Lenses

describeLoggingStatus_clusterIdentifier :: Lens' DescribeLoggingStatus Text

The identifier of the cluster from which to get the logging status.

Example: examplecluster

Destructuring the Response

data LoggingStatus

Describes the status of logging for a cluster.

See: newLoggingStatus smart constructor.

Constructors

LoggingStatus'

Fields

lastFailureTime :: Maybe ISO8601

The last time when logs failed to be delivered.

lastSuccessfulDeliveryTime :: Maybe ISO8601

The last time that logs were delivered.

s3KeyPrefix :: Maybe Text

The prefix applied to the log file names.

bucketName :: Maybe Text

The name of the S3 bucket where the log files are stored.

loggingEnabled :: Maybe Bool

true if logging is on, false if logging is off.

lastFailureMessage :: Maybe Text

The message indicating that logs failed to be delivered.

newLoggingStatus :: LoggingStatus

Create a value of LoggingStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastFailureTime:LoggingStatus', loggingStatus_lastFailureTime - The last time when logs failed to be delivered.

$sel:lastSuccessfulDeliveryTime:LoggingStatus', loggingStatus_lastSuccessfulDeliveryTime - The last time that logs were delivered.

$sel:s3KeyPrefix:LoggingStatus', loggingStatus_s3KeyPrefix - The prefix applied to the log file names.

$sel:bucketName:LoggingStatus', loggingStatus_bucketName - The name of the S3 bucket where the log files are stored.

$sel:loggingEnabled:LoggingStatus', loggingStatus_loggingEnabled - true if logging is on, false if logging is off.

$sel:lastFailureMessage:LoggingStatus', loggingStatus_lastFailureMessage - The message indicating that logs failed to be delivered.

Response Lenses

loggingStatus_lastFailureTime :: Lens' LoggingStatus (Maybe UTCTime)

The last time when logs failed to be delivered.

loggingStatus_lastSuccessfulDeliveryTime :: Lens' LoggingStatus (Maybe UTCTime)

The last time that logs were delivered.

loggingStatus_s3KeyPrefix :: Lens' LoggingStatus (Maybe Text)

The prefix applied to the log file names.

loggingStatus_bucketName :: Lens' LoggingStatus (Maybe Text)

The name of the S3 bucket where the log files are stored.

loggingStatus_loggingEnabled :: Lens' LoggingStatus (Maybe Bool)

true if logging is on, false if logging is off.

loggingStatus_lastFailureMessage :: Lens' LoggingStatus (Maybe Text)

The message indicating that logs failed to be delivered.