Amazonka.Redshift.DescribeUsageLimits

Description

Shows usage limits on a cluster. Results are filtered based on the combination of input usage limit identifier, cluster identifier, and feature type parameters:

If usage limit identifier, cluster identifier, and feature type are not provided, then all usage limit objects for the current account in the current region are returned.

If usage limit identifier is provided, then the corresponding usage limit object is returned.

If cluster identifier is provided, then all usage limit objects for the specified cluster are returned.

If cluster identifier and feature type are provided, then all usage limit objects for the combination of cluster and feature are returned.

This operation returns paginated results.

Creating a Request

data DescribeUsageLimits

See: newDescribeUsageLimits smart constructor.

Constructors

DescribeUsageLimits'

Fields

tagValues :: Maybe [Text]

A tag value or values for which you want to return all matching usage limit objects that are associated with the specified tag value or values. For example, suppose that you have parameter groups that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the usage limit objects that have either or both of these tag values associated with them.

usageLimitId :: Maybe Text

The identifier of the usage limit to describe.

tagKeys :: Maybe [Text]

A tag key or keys for which you want to return all matching usage limit objects that are associated with the specified key or keys. For example, suppose that you have parameter groups that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the usage limit objects have either or both of these tag keys associated with them.

clusterIdentifier :: Maybe Text

The identifier of the cluster for which you want to describe usage limits.

featureType :: Maybe UsageLimitFeatureType

The feature type for which you want to describe usage limits.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeUsageLimits request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

newDescribeUsageLimits :: DescribeUsageLimits

Create a value of DescribeUsageLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagValues:DescribeUsageLimits', describeUsageLimits_tagValues - A tag value or values for which you want to return all matching usage limit objects that are associated with the specified tag value or values. For example, suppose that you have parameter groups that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the usage limit objects that have either or both of these tag values associated with them.

$sel:usageLimitId:DescribeUsageLimits', describeUsageLimits_usageLimitId - The identifier of the usage limit to describe.

$sel:tagKeys:DescribeUsageLimits', describeUsageLimits_tagKeys - A tag key or keys for which you want to return all matching usage limit objects that are associated with the specified key or keys. For example, suppose that you have parameter groups that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the usage limit objects have either or both of these tag keys associated with them.

$sel:clusterIdentifier:DescribeUsageLimits', describeUsageLimits_clusterIdentifier - The identifier of the cluster for which you want to describe usage limits.

$sel:featureType:DescribeUsageLimits', describeUsageLimits_featureType - The feature type for which you want to describe usage limits.

$sel:marker:DescribeUsageLimits', describeUsageLimits_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeUsageLimits request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:maxRecords:DescribeUsageLimits', describeUsageLimits_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

Request Lenses

describeUsageLimits_tagValues :: Lens' DescribeUsageLimits (Maybe [Text])

A tag value or values for which you want to return all matching usage limit objects that are associated with the specified tag value or values. For example, suppose that you have parameter groups that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the usage limit objects that have either or both of these tag values associated with them.

describeUsageLimits_usageLimitId :: Lens' DescribeUsageLimits (Maybe Text)

The identifier of the usage limit to describe.

describeUsageLimits_tagKeys :: Lens' DescribeUsageLimits (Maybe [Text])

A tag key or keys for which you want to return all matching usage limit objects that are associated with the specified key or keys. For example, suppose that you have parameter groups that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the usage limit objects have either or both of these tag keys associated with them.

describeUsageLimits_clusterIdentifier :: Lens' DescribeUsageLimits (Maybe Text)

The identifier of the cluster for which you want to describe usage limits.

describeUsageLimits_featureType :: Lens' DescribeUsageLimits (Maybe UsageLimitFeatureType)

The feature type for which you want to describe usage limits.

describeUsageLimits_marker :: Lens' DescribeUsageLimits (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeUsageLimits request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeUsageLimits_maxRecords :: Lens' DescribeUsageLimits (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

Destructuring the Response

data DescribeUsageLimitsResponse

See: newDescribeUsageLimitsResponse smart constructor.

Constructors

DescribeUsageLimitsResponse'

Fields

usageLimits :: Maybe [UsageLimit]

Contains the output from the DescribeUsageLimits action.

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

httpStatus :: Int

The response's http status code.

newDescribeUsageLimitsResponse

Arguments

:: Int
$sel:httpStatus:DescribeUsageLimitsResponse'

-> DescribeUsageLimitsResponse

Create a value of DescribeUsageLimitsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:usageLimits:DescribeUsageLimitsResponse', describeUsageLimitsResponse_usageLimits - Contains the output from the DescribeUsageLimits action.

$sel:marker:DescribeUsageLimits', describeUsageLimitsResponse_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

$sel:httpStatus:DescribeUsageLimitsResponse', describeUsageLimitsResponse_httpStatus - The response's http status code.

Response Lenses

describeUsageLimitsResponse_usageLimits :: Lens' DescribeUsageLimitsResponse (Maybe [UsageLimit])

Contains the output from the DescribeUsageLimits action.

describeUsageLimitsResponse_marker :: Lens' DescribeUsageLimitsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

describeUsageLimitsResponse_httpStatus :: Lens' DescribeUsageLimitsResponse Int

The response's http status code.