Amazonka.Redshift.ModifySnapshotCopyRetentionPeriod

Description

Modifies the number of days to retain snapshots in the destination Amazon Web Services Region after they are copied from the source Amazon Web Services Region. By default, this operation only changes the retention period of copied automated snapshots. The retention periods for both new and existing copied automated snapshots are updated with the new retention period. You can set the manual option to change only the retention periods of copied manual snapshots. If you set this option, only newly copied manual snapshots have the new retention period.

Creating a Request

data ModifySnapshotCopyRetentionPeriod

See: newModifySnapshotCopyRetentionPeriod smart constructor.

Constructors

ModifySnapshotCopyRetentionPeriod'

Fields

manual :: Maybe Bool

Indicates whether to apply the snapshot retention period to newly copied manual snapshots instead of automated snapshots.

clusterIdentifier :: Text

The unique identifier of the cluster for which you want to change the retention period for either automated or manual snapshots that are copied to a destination Amazon Web Services Region.

Constraints: Must be the valid name of an existing cluster that has cross-region snapshot copy enabled.

retentionPeriod :: Int

The number of days to retain automated snapshots in the destination Amazon Web Services Region after they are copied from the source Amazon Web Services Region.

By default, this only changes the retention period of copied automated snapshots.

If you decrease the retention period for automated snapshots that are copied to a destination Amazon Web Services Region, Amazon Redshift deletes any existing automated snapshots that were copied to the destination Amazon Web Services Region and that fall outside of the new retention period.

Constraints: Must be at least 1 and no more than 35 for automated snapshots.

If you specify the manual option, only newly copied manual snapshots will have the new retention period.

If you specify the value of -1 newly copied manual snapshots are retained indefinitely.

Constraints: The number of days must be either -1 or an integer between 1 and 3,653 for manual snapshots.

newModifySnapshotCopyRetentionPeriod

Arguments

:: Text
$sel:clusterIdentifier:ModifySnapshotCopyRetentionPeriod'

-> Int
$sel:retentionPeriod:ModifySnapshotCopyRetentionPeriod'

-> ModifySnapshotCopyRetentionPeriod

Create a value of ModifySnapshotCopyRetentionPeriod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manual:ModifySnapshotCopyRetentionPeriod', modifySnapshotCopyRetentionPeriod_manual - Indicates whether to apply the snapshot retention period to newly copied manual snapshots instead of automated snapshots.

$sel:clusterIdentifier:ModifySnapshotCopyRetentionPeriod', modifySnapshotCopyRetentionPeriod_clusterIdentifier - The unique identifier of the cluster for which you want to change the retention period for either automated or manual snapshots that are copied to a destination Amazon Web Services Region.

Constraints: Must be the valid name of an existing cluster that has cross-region snapshot copy enabled.

$sel:retentionPeriod:ModifySnapshotCopyRetentionPeriod', modifySnapshotCopyRetentionPeriod_retentionPeriod - The number of days to retain automated snapshots in the destination Amazon Web Services Region after they are copied from the source Amazon Web Services Region.

By default, this only changes the retention period of copied automated snapshots.

If you decrease the retention period for automated snapshots that are copied to a destination Amazon Web Services Region, Amazon Redshift deletes any existing automated snapshots that were copied to the destination Amazon Web Services Region and that fall outside of the new retention period.

Constraints: Must be at least 1 and no more than 35 for automated snapshots.

If you specify the manual option, only newly copied manual snapshots will have the new retention period.

If you specify the value of -1 newly copied manual snapshots are retained indefinitely.

Constraints: The number of days must be either -1 or an integer between 1 and 3,653 for manual snapshots.

Request Lenses

modifySnapshotCopyRetentionPeriod_manual :: Lens' ModifySnapshotCopyRetentionPeriod (Maybe Bool)

Indicates whether to apply the snapshot retention period to newly copied manual snapshots instead of automated snapshots.

modifySnapshotCopyRetentionPeriod_clusterIdentifier :: Lens' ModifySnapshotCopyRetentionPeriod Text

The unique identifier of the cluster for which you want to change the retention period for either automated or manual snapshots that are copied to a destination Amazon Web Services Region.

Constraints: Must be the valid name of an existing cluster that has cross-region snapshot copy enabled.

modifySnapshotCopyRetentionPeriod_retentionPeriod :: Lens' ModifySnapshotCopyRetentionPeriod Int

The number of days to retain automated snapshots in the destination Amazon Web Services Region after they are copied from the source Amazon Web Services Region.

By default, this only changes the retention period of copied automated snapshots.

If you decrease the retention period for automated snapshots that are copied to a destination Amazon Web Services Region, Amazon Redshift deletes any existing automated snapshots that were copied to the destination Amazon Web Services Region and that fall outside of the new retention period.

Constraints: Must be at least 1 and no more than 35 for automated snapshots.

If you specify the manual option, only newly copied manual snapshots will have the new retention period.

If you specify the value of -1 newly copied manual snapshots are retained indefinitely.

Constraints: The number of days must be either -1 or an integer between 1 and 3,653 for manual snapshots.

Destructuring the Response

data ModifySnapshotCopyRetentionPeriodResponse

See: newModifySnapshotCopyRetentionPeriodResponse smart constructor.

Constructors

ModifySnapshotCopyRetentionPeriodResponse'

Fields

cluster :: Maybe Cluster

httpStatus :: Int

The response's http status code.

newModifySnapshotCopyRetentionPeriodResponse

Arguments

:: Int
$sel:httpStatus:ModifySnapshotCopyRetentionPeriodResponse'

-> ModifySnapshotCopyRetentionPeriodResponse

Create a value of ModifySnapshotCopyRetentionPeriodResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:ModifySnapshotCopyRetentionPeriodResponse', modifySnapshotCopyRetentionPeriodResponse_cluster - Undocumented member.

$sel:httpStatus:ModifySnapshotCopyRetentionPeriodResponse', modifySnapshotCopyRetentionPeriodResponse_httpStatus - The response's http status code.

Response Lenses

modifySnapshotCopyRetentionPeriodResponse_cluster :: Lens' ModifySnapshotCopyRetentionPeriodResponse (Maybe Cluster)

Undocumented member.

modifySnapshotCopyRetentionPeriodResponse_httpStatus :: Lens' ModifySnapshotCopyRetentionPeriodResponse Int

The response's http status code.