Amazonka.Redshift.Types.ClusterParameterGroupStatus

Description

Documentation

data ClusterParameterGroupStatus

Describes the status of a parameter group.

See: newClusterParameterGroupStatus smart constructor.

Constructors

ClusterParameterGroupStatus'

Fields

clusterParameterStatusList :: Maybe [ClusterParameterStatus]

The list of parameter statuses.

For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide.

parameterApplyStatus :: Maybe Text

The status of parameter updates.

parameterGroupName :: Maybe Text

The name of the cluster parameter group.

newClusterParameterGroupStatus :: ClusterParameterGroupStatus

Create a value of ClusterParameterGroupStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterParameterStatusList:ClusterParameterGroupStatus', clusterParameterGroupStatus_clusterParameterStatusList - The list of parameter statuses.

For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide.

$sel:parameterApplyStatus:ClusterParameterGroupStatus', clusterParameterGroupStatus_parameterApplyStatus - The status of parameter updates.

$sel:parameterGroupName:ClusterParameterGroupStatus', clusterParameterGroupStatus_parameterGroupName - The name of the cluster parameter group.

clusterParameterGroupStatus_clusterParameterStatusList :: Lens' ClusterParameterGroupStatus (Maybe [ClusterParameterStatus])

The list of parameter statuses.

For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide.

clusterParameterGroupStatus_parameterApplyStatus :: Lens' ClusterParameterGroupStatus (Maybe Text)

The status of parameter updates.

clusterParameterGroupStatus_parameterGroupName :: Lens' ClusterParameterGroupStatus (Maybe Text)

The name of the cluster parameter group.