Amazonka.Redshift.Types.DeferredMaintenanceWindow

Description

Documentation

data DeferredMaintenanceWindow

Describes a deferred maintenance window

See: newDeferredMaintenanceWindow smart constructor.

Constructors

DeferredMaintenanceWindow'

Fields

deferMaintenanceEndTime :: Maybe ISO8601

A timestamp for the end of the time period when we defer maintenance.

deferMaintenanceStartTime :: Maybe ISO8601

A timestamp for the beginning of the time period when we defer maintenance.

deferMaintenanceIdentifier :: Maybe Text

A unique identifier for the maintenance window.

newDeferredMaintenanceWindow :: DeferredMaintenanceWindow

Create a value of DeferredMaintenanceWindow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deferMaintenanceEndTime:DeferredMaintenanceWindow', deferredMaintenanceWindow_deferMaintenanceEndTime - A timestamp for the end of the time period when we defer maintenance.

$sel:deferMaintenanceStartTime:DeferredMaintenanceWindow', deferredMaintenanceWindow_deferMaintenanceStartTime - A timestamp for the beginning of the time period when we defer maintenance.

$sel:deferMaintenanceIdentifier:DeferredMaintenanceWindow', deferredMaintenanceWindow_deferMaintenanceIdentifier - A unique identifier for the maintenance window.

deferredMaintenanceWindow_deferMaintenanceEndTime :: Lens' DeferredMaintenanceWindow (Maybe UTCTime)

A timestamp for the end of the time period when we defer maintenance.

deferredMaintenanceWindow_deferMaintenanceStartTime :: Lens' DeferredMaintenanceWindow (Maybe UTCTime)

A timestamp for the beginning of the time period when we defer maintenance.

deferredMaintenanceWindow_deferMaintenanceIdentifier :: Lens' DeferredMaintenanceWindow (Maybe Text)

A unique identifier for the maintenance window.