Amazonka.Redshift.Types.ReservedNodeOffering

Description

Documentation

data ReservedNodeOffering

Describes a reserved node offering.

See: newReservedNodeOffering smart constructor.

Constructors

ReservedNodeOffering'

Fields

reservedNodeOfferingType :: Maybe ReservedNodeOfferingType

currencyCode :: Maybe Text

The currency code for the compute nodes offering.

reservedNodeOfferingId :: Maybe Text

The offering identifier.

recurringCharges :: Maybe [RecurringCharge]

The charge to your account regardless of whether you are creating any clusters using the node offering. Recurring charges are only in effect for heavy-utilization reserved nodes.

offeringType :: Maybe Text

The anticipated utilization of the reserved node, as defined in the reserved node offering.

usagePrice :: Maybe Double

The rate you are charged for each hour the cluster that is using the offering is running.

nodeType :: Maybe Text

The node type offered by the reserved node offering.

fixedPrice :: Maybe Double

The upfront fixed charge you will pay to purchase the specific reserved node offering.

duration :: Maybe Int

The duration, in seconds, for which the offering will reserve the node.

newReservedNodeOffering :: ReservedNodeOffering

Create a value of ReservedNodeOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedNodeOfferingType:ReservedNodeOffering', reservedNodeOffering_reservedNodeOfferingType -

$sel:currencyCode:ReservedNodeOffering', reservedNodeOffering_currencyCode - The currency code for the compute nodes offering.

$sel:reservedNodeOfferingId:ReservedNodeOffering', reservedNodeOffering_reservedNodeOfferingId - The offering identifier.

$sel:recurringCharges:ReservedNodeOffering', reservedNodeOffering_recurringCharges - The charge to your account regardless of whether you are creating any clusters using the node offering. Recurring charges are only in effect for heavy-utilization reserved nodes.

$sel:offeringType:ReservedNodeOffering', reservedNodeOffering_offeringType - The anticipated utilization of the reserved node, as defined in the reserved node offering.

$sel:usagePrice:ReservedNodeOffering', reservedNodeOffering_usagePrice - The rate you are charged for each hour the cluster that is using the offering is running.

$sel:nodeType:ReservedNodeOffering', reservedNodeOffering_nodeType - The node type offered by the reserved node offering.

$sel:fixedPrice:ReservedNodeOffering', reservedNodeOffering_fixedPrice - The upfront fixed charge you will pay to purchase the specific reserved node offering.

$sel:duration:ReservedNodeOffering', reservedNodeOffering_duration - The duration, in seconds, for which the offering will reserve the node.

reservedNodeOffering_reservedNodeOfferingType :: Lens' ReservedNodeOffering (Maybe ReservedNodeOfferingType)

reservedNodeOffering_currencyCode :: Lens' ReservedNodeOffering (Maybe Text)

The currency code for the compute nodes offering.

reservedNodeOffering_reservedNodeOfferingId :: Lens' ReservedNodeOffering (Maybe Text)

The offering identifier.

reservedNodeOffering_recurringCharges :: Lens' ReservedNodeOffering (Maybe [RecurringCharge])

The charge to your account regardless of whether you are creating any clusters using the node offering. Recurring charges are only in effect for heavy-utilization reserved nodes.

reservedNodeOffering_offeringType :: Lens' ReservedNodeOffering (Maybe Text)

The anticipated utilization of the reserved node, as defined in the reserved node offering.

reservedNodeOffering_usagePrice :: Lens' ReservedNodeOffering (Maybe Double)

The rate you are charged for each hour the cluster that is using the offering is running.

reservedNodeOffering_nodeType :: Lens' ReservedNodeOffering (Maybe Text)

The node type offered by the reserved node offering.

reservedNodeOffering_fixedPrice :: Lens' ReservedNodeOffering (Maybe Double)

The upfront fixed charge you will pay to purchase the specific reserved node offering.

reservedNodeOffering_duration :: Lens' ReservedNodeOffering (Maybe Int)

The duration, in seconds, for which the offering will reserve the node.