Amazonka.Redshift.Types.Snapshot

Description

Documentation

data Snapshot

Describes a snapshot.

See: newSnapshot smart constructor.

Constructors

Snapshot'

Fields

status :: Maybe Text

The snapshot status. The value of the status depends on the API operation used:

CreateClusterSnapshot and CopyClusterSnapshot returns status as "creating".

DescribeClusterSnapshots returns status as "creating", "available", "final snapshot", or "failed".

DeleteClusterSnapshot returns status as "deleted".

restorableNodeTypes :: Maybe [Text]

The list of node types that this cluster snapshot is able to restore into.

accountsWithRestoreAccess :: Maybe [AccountWithRestoreAccess]

A list of the Amazon Web Services accounts authorized to restore the snapshot. Returns null if no accounts are authorized. Visible only to the snapshot owner.

manualSnapshotRetentionPeriod :: Maybe Int

The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

enhancedVpcRouting :: Maybe Bool

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

snapshotIdentifier :: Maybe Text

The snapshot identifier that is provided in the request.

engineFullVersion :: Maybe Text

The cluster version of the cluster used to create the snapshot. For example, 1.0.15503.

encryptedWithHSM :: Maybe Bool

A boolean that indicates whether the snapshot data is encrypted using the HSM keys of the source cluster. true indicates that the data is encrypted using HSM keys.

masterUsername :: Maybe Text

The admin user name for the cluster.

sourceRegion :: Maybe Text

The source region from which the snapshot was copied.

maintenanceTrackName :: Maybe Text

The name of the maintenance track for the snapshot.

snapshotRetentionStartTime :: Maybe ISO8601

A timestamp representing the start of the retention period for the snapshot.

manualSnapshotRemainingDays :: Maybe Int

The number of days until a manual snapshot will pass its retention period.

vpcId :: Maybe Text

The VPC identifier of the cluster if the snapshot is from a cluster in a VPC. Otherwise, this field is not in the output.

backupProgressInMegaBytes :: Maybe Double

The number of megabytes that have been transferred to the snapshot backup.

encrypted :: Maybe Bool

If true, the data in the snapshot is encrypted at rest.

clusterIdentifier :: Maybe Text

The identifier of the cluster for which the snapshot was taken.

numberOfNodes :: Maybe Int

The number of nodes in the cluster.

snapshotType :: Maybe Text

The snapshot type. Snapshots created using CreateClusterSnapshot and CopyClusterSnapshot are of type "manual".

kmsKeyId :: Maybe Text

The Key Management Service (KMS) key ID of the encryption key that was used to encrypt data in the cluster from which the snapshot was taken.

availabilityZone :: Maybe Text

The Availability Zone in which the cluster was created.

currentBackupRateInMegaBytesPerSecond :: Maybe Double

The number of megabytes per second being transferred to the snapshot backup. Returns 0 for a completed backup.

snapshotCreateTime :: Maybe ISO8601

The time (in UTC format) when Amazon Redshift began the snapshot. A snapshot contains a copy of the cluster data as of this exact time.

clusterVersion :: Maybe Text

The version ID of the Amazon Redshift engine that is running on the cluster.

ownerAccount :: Maybe Text

For manual snapshots, the Amazon Web Services account used to create or copy the snapshot. For automatic snapshots, the owner of the cluster. The owner can perform all snapshot actions, such as sharing a manual snapshot.

nodeType :: Maybe Text

The node type of the nodes in the cluster.

elapsedTimeInSeconds :: Maybe Integer

The amount of time an in-progress snapshot backup has been running, or the amount of time it took a completed backup to finish.

clusterCreateTime :: Maybe ISO8601

The time (UTC) when the cluster was originally created.

estimatedSecondsToCompletion :: Maybe Integer

The estimate of the time remaining before the snapshot backup will complete. Returns 0 for a completed backup.

actualIncrementalBackupSizeInMegaBytes :: Maybe Double

The size of the incremental backup.

tags :: Maybe [Tag]

The list of tags for the cluster snapshot.

port :: Maybe Int

The port that the cluster is listening on.

totalBackupSizeInMegaBytes :: Maybe Double

The size of the complete set of backup data that would be used to restore the cluster.

dbName :: Maybe Text

The name of the database that was created when the cluster was created.

newSnapshot :: Snapshot

Create a value of Snapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Snapshot', snapshot_status - The snapshot status. The value of the status depends on the API operation used:

CreateClusterSnapshot and CopyClusterSnapshot returns status as "creating".

DescribeClusterSnapshots returns status as "creating", "available", "final snapshot", or "failed".

DeleteClusterSnapshot returns status as "deleted".

$sel:restorableNodeTypes:Snapshot', snapshot_restorableNodeTypes - The list of node types that this cluster snapshot is able to restore into.

$sel:accountsWithRestoreAccess:Snapshot', snapshot_accountsWithRestoreAccess - A list of the Amazon Web Services accounts authorized to restore the snapshot. Returns null if no accounts are authorized. Visible only to the snapshot owner.

$sel:manualSnapshotRetentionPeriod:Snapshot', snapshot_manualSnapshotRetentionPeriod - The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

$sel:enhancedVpcRouting:Snapshot', snapshot_enhancedVpcRouting - An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

$sel:snapshotIdentifier:Snapshot', snapshot_snapshotIdentifier - The snapshot identifier that is provided in the request.

$sel:engineFullVersion:Snapshot', snapshot_engineFullVersion - The cluster version of the cluster used to create the snapshot. For example, 1.0.15503.

$sel:encryptedWithHSM:Snapshot', snapshot_encryptedWithHSM - A boolean that indicates whether the snapshot data is encrypted using the HSM keys of the source cluster. true indicates that the data is encrypted using HSM keys.

$sel:masterUsername:Snapshot', snapshot_masterUsername - The admin user name for the cluster.

$sel:sourceRegion:Snapshot', snapshot_sourceRegion - The source region from which the snapshot was copied.

$sel:maintenanceTrackName:Snapshot', snapshot_maintenanceTrackName - The name of the maintenance track for the snapshot.

$sel:snapshotRetentionStartTime:Snapshot', snapshot_snapshotRetentionStartTime - A timestamp representing the start of the retention period for the snapshot.

$sel:manualSnapshotRemainingDays:Snapshot', snapshot_manualSnapshotRemainingDays - The number of days until a manual snapshot will pass its retention period.

$sel:vpcId:Snapshot', snapshot_vpcId - The VPC identifier of the cluster if the snapshot is from a cluster in a VPC. Otherwise, this field is not in the output.

$sel:backupProgressInMegaBytes:Snapshot', snapshot_backupProgressInMegaBytes - The number of megabytes that have been transferred to the snapshot backup.

$sel:encrypted:Snapshot', snapshot_encrypted - If true, the data in the snapshot is encrypted at rest.

$sel:clusterIdentifier:Snapshot', snapshot_clusterIdentifier - The identifier of the cluster for which the snapshot was taken.

$sel:numberOfNodes:Snapshot', snapshot_numberOfNodes - The number of nodes in the cluster.

$sel:snapshotType:Snapshot', snapshot_snapshotType - The snapshot type. Snapshots created using CreateClusterSnapshot and CopyClusterSnapshot are of type "manual".

$sel:kmsKeyId:Snapshot', snapshot_kmsKeyId - The Key Management Service (KMS) key ID of the encryption key that was used to encrypt data in the cluster from which the snapshot was taken.

$sel:availabilityZone:Snapshot', snapshot_availabilityZone - The Availability Zone in which the cluster was created.

$sel:currentBackupRateInMegaBytesPerSecond:Snapshot', snapshot_currentBackupRateInMegaBytesPerSecond - The number of megabytes per second being transferred to the snapshot backup. Returns 0 for a completed backup.

$sel:snapshotCreateTime:Snapshot', snapshot_snapshotCreateTime - The time (in UTC format) when Amazon Redshift began the snapshot. A snapshot contains a copy of the cluster data as of this exact time.

$sel:clusterVersion:Snapshot', snapshot_clusterVersion - The version ID of the Amazon Redshift engine that is running on the cluster.

$sel:ownerAccount:Snapshot', snapshot_ownerAccount - For manual snapshots, the Amazon Web Services account used to create or copy the snapshot. For automatic snapshots, the owner of the cluster. The owner can perform all snapshot actions, such as sharing a manual snapshot.

$sel:nodeType:Snapshot', snapshot_nodeType - The node type of the nodes in the cluster.

$sel:elapsedTimeInSeconds:Snapshot', snapshot_elapsedTimeInSeconds - The amount of time an in-progress snapshot backup has been running, or the amount of time it took a completed backup to finish.

$sel:clusterCreateTime:Snapshot', snapshot_clusterCreateTime - The time (UTC) when the cluster was originally created.

$sel:estimatedSecondsToCompletion:Snapshot', snapshot_estimatedSecondsToCompletion - The estimate of the time remaining before the snapshot backup will complete. Returns 0 for a completed backup.

$sel:actualIncrementalBackupSizeInMegaBytes:Snapshot', snapshot_actualIncrementalBackupSizeInMegaBytes - The size of the incremental backup.

$sel:tags:Snapshot', snapshot_tags - The list of tags for the cluster snapshot.

$sel:port:Snapshot', snapshot_port - The port that the cluster is listening on.

$sel:totalBackupSizeInMegaBytes:Snapshot', snapshot_totalBackupSizeInMegaBytes - The size of the complete set of backup data that would be used to restore the cluster.

$sel:dbName:Snapshot', snapshot_dbName - The name of the database that was created when the cluster was created.

snapshot_status :: Lens' Snapshot (Maybe Text)

The snapshot status. The value of the status depends on the API operation used:

CreateClusterSnapshot and CopyClusterSnapshot returns status as "creating".

DescribeClusterSnapshots returns status as "creating", "available", "final snapshot", or "failed".

DeleteClusterSnapshot returns status as "deleted".

snapshot_restorableNodeTypes :: Lens' Snapshot (Maybe [Text])

The list of node types that this cluster snapshot is able to restore into.

snapshot_accountsWithRestoreAccess :: Lens' Snapshot (Maybe [AccountWithRestoreAccess])

A list of the Amazon Web Services accounts authorized to restore the snapshot. Returns null if no accounts are authorized. Visible only to the snapshot owner.

snapshot_manualSnapshotRetentionPeriod :: Lens' Snapshot (Maybe Int)

The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

snapshot_enhancedVpcRouting :: Lens' Snapshot (Maybe Bool)

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

snapshot_snapshotIdentifier :: Lens' Snapshot (Maybe Text)

The snapshot identifier that is provided in the request.

snapshot_engineFullVersion :: Lens' Snapshot (Maybe Text)

The cluster version of the cluster used to create the snapshot. For example, 1.0.15503.

snapshot_encryptedWithHSM :: Lens' Snapshot (Maybe Bool)

A boolean that indicates whether the snapshot data is encrypted using the HSM keys of the source cluster. true indicates that the data is encrypted using HSM keys.

snapshot_masterUsername :: Lens' Snapshot (Maybe Text)

The admin user name for the cluster.

snapshot_sourceRegion :: Lens' Snapshot (Maybe Text)

The source region from which the snapshot was copied.

snapshot_maintenanceTrackName :: Lens' Snapshot (Maybe Text)

The name of the maintenance track for the snapshot.

snapshot_snapshotRetentionStartTime :: Lens' Snapshot (Maybe UTCTime)

A timestamp representing the start of the retention period for the snapshot.

snapshot_manualSnapshotRemainingDays :: Lens' Snapshot (Maybe Int)

The number of days until a manual snapshot will pass its retention period.

snapshot_vpcId :: Lens' Snapshot (Maybe Text)

The VPC identifier of the cluster if the snapshot is from a cluster in a VPC. Otherwise, this field is not in the output.

snapshot_backupProgressInMegaBytes :: Lens' Snapshot (Maybe Double)

The number of megabytes that have been transferred to the snapshot backup.

snapshot_encrypted :: Lens' Snapshot (Maybe Bool)

If true, the data in the snapshot is encrypted at rest.

snapshot_clusterIdentifier :: Lens' Snapshot (Maybe Text)

The identifier of the cluster for which the snapshot was taken.

snapshot_numberOfNodes :: Lens' Snapshot (Maybe Int)

The number of nodes in the cluster.

snapshot_snapshotType :: Lens' Snapshot (Maybe Text)

The snapshot type. Snapshots created using CreateClusterSnapshot and CopyClusterSnapshot are of type "manual".

snapshot_kmsKeyId :: Lens' Snapshot (Maybe Text)

The Key Management Service (KMS) key ID of the encryption key that was used to encrypt data in the cluster from which the snapshot was taken.

snapshot_availabilityZone :: Lens' Snapshot (Maybe Text)

The Availability Zone in which the cluster was created.

snapshot_currentBackupRateInMegaBytesPerSecond :: Lens' Snapshot (Maybe Double)

The number of megabytes per second being transferred to the snapshot backup. Returns 0 for a completed backup.

snapshot_snapshotCreateTime :: Lens' Snapshot (Maybe UTCTime)

The time (in UTC format) when Amazon Redshift began the snapshot. A snapshot contains a copy of the cluster data as of this exact time.

snapshot_clusterVersion :: Lens' Snapshot (Maybe Text)

The version ID of the Amazon Redshift engine that is running on the cluster.

snapshot_ownerAccount :: Lens' Snapshot (Maybe Text)

For manual snapshots, the Amazon Web Services account used to create or copy the snapshot. For automatic snapshots, the owner of the cluster. The owner can perform all snapshot actions, such as sharing a manual snapshot.

snapshot_nodeType :: Lens' Snapshot (Maybe Text)

The node type of the nodes in the cluster.

snapshot_elapsedTimeInSeconds :: Lens' Snapshot (Maybe Integer)

The amount of time an in-progress snapshot backup has been running, or the amount of time it took a completed backup to finish.

snapshot_clusterCreateTime :: Lens' Snapshot (Maybe UTCTime)

The time (UTC) when the cluster was originally created.

snapshot_estimatedSecondsToCompletion :: Lens' Snapshot (Maybe Integer)

The estimate of the time remaining before the snapshot backup will complete. Returns 0 for a completed backup.

snapshot_actualIncrementalBackupSizeInMegaBytes :: Lens' Snapshot (Maybe Double)

The size of the incremental backup.

snapshot_tags :: Lens' Snapshot (Maybe [Tag])

The list of tags for the cluster snapshot.

snapshot_port :: Lens' Snapshot (Maybe Int)

The port that the cluster is listening on.

snapshot_totalBackupSizeInMegaBytes :: Lens' Snapshot (Maybe Double)

The size of the complete set of backup data that would be used to restore the cluster.

snapshot_dbName :: Lens' Snapshot (Maybe Text)

The name of the database that was created when the cluster was created.