Amazonka.RedshiftData.DescribeTable

Description

Describes the detailed information about a table from metadata in the cluster. The information includes its columns. A token is returned to page through the column list. Depending on the authorization method, use one of the following combinations of request parameters:

Secrets Manager - specify the Amazon Resource Name (ARN) of the secret, the database name, and the cluster identifier that matches the cluster in the secret.

Temporary credentials - specify the cluster identifier, the database name, and the database user name. Permission to call the redshift:GetClusterCredentials operation is required to use this method.

This operation returns paginated results.

Creating a Request

data DescribeTable

See: newDescribeTable smart constructor.

Constructors

DescribeTable'

Fields

dbUser :: Maybe Text

The database user name. This parameter is required when authenticating using temporary credentials.

connectedDatabase :: Maybe Text

A database name. The connected database is specified when you connect with your authentication credentials.

schema :: Maybe Text

The schema that contains the table. If no schema is specified, then matching tables for all schemas are returned.

nextToken :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned NextToken value in the next NextToken parameter and retrying the command. If the NextToken field is empty, all response records have been retrieved for the request.

secretArn :: Maybe Text

The name or ARN of the secret that enables access to the database. This parameter is required when authenticating using Secrets Manager.

table :: Maybe Text

The table name. If no table is specified, then all tables for all matching schemas are returned. If no table and no schema is specified, then all tables for all schemas in the database are returned

maxResults :: Maybe Natural

The maximum number of tables to return in the response. If more tables exist than fit in one response, then NextToken is returned to page through the results.

clusterIdentifier :: Text

The cluster identifier. This parameter is required when authenticating using either Secrets Manager or temporary credentials.

database :: Text

The name of the database that contains the tables to be described. If ConnectedDatabase is not specified, this is also the database to connect to with your authentication credentials.

newDescribeTable

Arguments

:: Text
$sel:clusterIdentifier:DescribeTable'

-> Text
$sel:database:DescribeTable'

-> DescribeTable

Create a value of DescribeTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbUser:DescribeTable', describeTable_dbUser - The database user name. This parameter is required when authenticating using temporary credentials.

$sel:connectedDatabase:DescribeTable', describeTable_connectedDatabase - A database name. The connected database is specified when you connect with your authentication credentials.

$sel:schema:DescribeTable', describeTable_schema - The schema that contains the table. If no schema is specified, then matching tables for all schemas are returned.

$sel:nextToken:DescribeTable', describeTable_nextToken - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned NextToken value in the next NextToken parameter and retrying the command. If the NextToken field is empty, all response records have been retrieved for the request.

$sel:secretArn:DescribeTable', describeTable_secretArn - The name or ARN of the secret that enables access to the database. This parameter is required when authenticating using Secrets Manager.

$sel:table:DescribeTable', describeTable_table - The table name. If no table is specified, then all tables for all matching schemas are returned. If no table and no schema is specified, then all tables for all schemas in the database are returned

$sel:maxResults:DescribeTable', describeTable_maxResults - The maximum number of tables to return in the response. If more tables exist than fit in one response, then NextToken is returned to page through the results.

$sel:clusterIdentifier:DescribeTable', describeTable_clusterIdentifier - The cluster identifier. This parameter is required when authenticating using either Secrets Manager or temporary credentials.

$sel:database:DescribeTable', describeTable_database - The name of the database that contains the tables to be described. If ConnectedDatabase is not specified, this is also the database to connect to with your authentication credentials.

Request Lenses

describeTable_dbUser :: Lens' DescribeTable (Maybe Text)

The database user name. This parameter is required when authenticating using temporary credentials.

describeTable_connectedDatabase :: Lens' DescribeTable (Maybe Text)

A database name. The connected database is specified when you connect with your authentication credentials.

describeTable_schema :: Lens' DescribeTable (Maybe Text)

The schema that contains the table. If no schema is specified, then matching tables for all schemas are returned.

describeTable_nextToken :: Lens' DescribeTable (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned NextToken value in the next NextToken parameter and retrying the command. If the NextToken field is empty, all response records have been retrieved for the request.

describeTable_secretArn :: Lens' DescribeTable (Maybe Text)

The name or ARN of the secret that enables access to the database. This parameter is required when authenticating using Secrets Manager.

describeTable_table :: Lens' DescribeTable (Maybe Text)

The table name. If no table is specified, then all tables for all matching schemas are returned. If no table and no schema is specified, then all tables for all schemas in the database are returned

describeTable_maxResults :: Lens' DescribeTable (Maybe Natural)

The maximum number of tables to return in the response. If more tables exist than fit in one response, then NextToken is returned to page through the results.

describeTable_clusterIdentifier :: Lens' DescribeTable Text

The cluster identifier. This parameter is required when authenticating using either Secrets Manager or temporary credentials.

describeTable_database :: Lens' DescribeTable Text

The name of the database that contains the tables to be described. If ConnectedDatabase is not specified, this is also the database to connect to with your authentication credentials.

Destructuring the Response

data DescribeTableResponse

See: newDescribeTableResponse smart constructor.

Constructors

DescribeTableResponse'

Fields

columnList :: Maybe [ColumnMetadata]

A list of columns in the table.

nextToken :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned NextToken value in the next NextToken parameter and retrying the command. If the NextToken field is empty, all response records have been retrieved for the request.

tableName :: Maybe Text

The table name.

httpStatus :: Int

The response's http status code.

newDescribeTableResponse

Arguments

:: Int
$sel:httpStatus:DescribeTableResponse'

-> DescribeTableResponse

Create a value of DescribeTableResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:columnList:DescribeTableResponse', describeTableResponse_columnList - A list of columns in the table.

$sel:nextToken:DescribeTable', describeTableResponse_nextToken - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned NextToken value in the next NextToken parameter and retrying the command. If the NextToken field is empty, all response records have been retrieved for the request.

$sel:tableName:DescribeTableResponse', describeTableResponse_tableName - The table name.

$sel:httpStatus:DescribeTableResponse', describeTableResponse_httpStatus - The response's http status code.

Response Lenses

describeTableResponse_columnList :: Lens' DescribeTableResponse (Maybe [ColumnMetadata])

A list of columns in the table.

describeTableResponse_nextToken :: Lens' DescribeTableResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned NextToken value in the next NextToken parameter and retrying the command. If the NextToken field is empty, all response records have been retrieved for the request.

describeTableResponse_tableName :: Lens' DescribeTableResponse (Maybe Text)

The table name.

describeTableResponse_httpStatus :: Lens' DescribeTableResponse Int

The response's http status code.