Amazonka.RedshiftData.Types

Description

Service Configuration

defaultService :: Service

API version 2019-12-20 of the Amazon Redshift Data API Service SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The Amazon Redshift Data API operation failed due to invalid input.

_ExecuteStatementException :: AsError a => Getting (First ServiceError) a ServiceError

The SQL statement encountered an environmental error while running.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The Amazon Redshift Data API operation failed due to invalid input.

_ActiveStatementsExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of active statements exceeds the limit.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The Amazon Redshift Data API operation failed due to a missing resource.

_BatchExecuteStatementException :: AsError a => Getting (First ServiceError) a ServiceError

An SQL statement encountered an environmental error while running.

StatementStatusString

newtype StatementStatusString

Constructors

StatementStatusString'

Fields

fromStatementStatusString :: Text

Bundled Patterns

pattern StatementStatusString_ABORTED :: StatementStatusString

pattern StatementStatusString_FAILED :: StatementStatusString

pattern StatementStatusString_FINISHED :: StatementStatusString

pattern StatementStatusString_PICKED :: StatementStatusString

pattern StatementStatusString_STARTED :: StatementStatusString

pattern StatementStatusString_SUBMITTED :: StatementStatusString

StatusString

newtype StatusString

Constructors

StatusString'

Fields

fromStatusString :: Text

Bundled Patterns

pattern StatusString_ABORTED :: StatusString

pattern StatusString_ALL :: StatusString

pattern StatusString_FAILED :: StatusString

pattern StatusString_FINISHED :: StatusString

pattern StatusString_PICKED :: StatusString

pattern StatusString_STARTED :: StatusString

pattern StatusString_SUBMITTED :: StatusString

ColumnMetadata

data ColumnMetadata

The properties (metadata) of a column.

See: newColumnMetadata smart constructor.

Constructors

ColumnMetadata'

Fields

length :: Maybe Int

The length of the column.

typeName :: Maybe Text

The database-specific data type of the column.

isCaseSensitive :: Maybe Bool

A value that indicates whether the column is case-sensitive.

columnDefault :: Maybe Text

The default value of the column.

isCurrency :: Maybe Bool

A value that indicates whether the column contains currency values.

scale :: Maybe Int

The scale value of a decimal number column.

precision :: Maybe Int

The precision value of a decimal number column.

schemaName :: Maybe Text

The name of the schema that contains the table that includes the column.

name :: Maybe Text

The name of the column.

isSigned :: Maybe Bool

A value that indicates whether an integer column is signed.

label :: Maybe Text

The label for the column.

nullable :: Maybe Int

A value that indicates whether the column is nullable.

tableName :: Maybe Text

The name of the table that includes the column.

newColumnMetadata :: ColumnMetadata

Create a value of ColumnMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:length:ColumnMetadata', columnMetadata_length - The length of the column.

$sel:typeName:ColumnMetadata', columnMetadata_typeName - The database-specific data type of the column.

$sel:isCaseSensitive:ColumnMetadata', columnMetadata_isCaseSensitive - A value that indicates whether the column is case-sensitive.

$sel:columnDefault:ColumnMetadata', columnMetadata_columnDefault - The default value of the column.

$sel:isCurrency:ColumnMetadata', columnMetadata_isCurrency - A value that indicates whether the column contains currency values.

$sel:scale:ColumnMetadata', columnMetadata_scale - The scale value of a decimal number column.

$sel:precision:ColumnMetadata', columnMetadata_precision - The precision value of a decimal number column.

$sel:schemaName:ColumnMetadata', columnMetadata_schemaName - The name of the schema that contains the table that includes the column.

$sel:name:ColumnMetadata', columnMetadata_name - The name of the column.

$sel:isSigned:ColumnMetadata', columnMetadata_isSigned - A value that indicates whether an integer column is signed.

$sel:label:ColumnMetadata', columnMetadata_label - The label for the column.

$sel:nullable:ColumnMetadata', columnMetadata_nullable - A value that indicates whether the column is nullable.

$sel:tableName:ColumnMetadata', columnMetadata_tableName - The name of the table that includes the column.

columnMetadata_length :: Lens' ColumnMetadata (Maybe Int)

The length of the column.

columnMetadata_typeName :: Lens' ColumnMetadata (Maybe Text)

The database-specific data type of the column.

columnMetadata_isCaseSensitive :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether the column is case-sensitive.

columnMetadata_columnDefault :: Lens' ColumnMetadata (Maybe Text)

The default value of the column.

columnMetadata_isCurrency :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether the column contains currency values.

columnMetadata_scale :: Lens' ColumnMetadata (Maybe Int)

The scale value of a decimal number column.

columnMetadata_precision :: Lens' ColumnMetadata (Maybe Int)

The precision value of a decimal number column.

columnMetadata_schemaName :: Lens' ColumnMetadata (Maybe Text)

The name of the schema that contains the table that includes the column.

columnMetadata_name :: Lens' ColumnMetadata (Maybe Text)

The name of the column.

columnMetadata_isSigned :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether an integer column is signed.

columnMetadata_label :: Lens' ColumnMetadata (Maybe Text)

The label for the column.

columnMetadata_nullable :: Lens' ColumnMetadata (Maybe Int)

A value that indicates whether the column is nullable.

columnMetadata_tableName :: Lens' ColumnMetadata (Maybe Text)

The name of the table that includes the column.

Field

data Field

A data value in a column.

See: newField smart constructor.

Constructors

Field'

Fields

doubleValue :: Maybe Double

A value of the double data type.

stringValue :: Maybe Text

A value of the string data type.

longValue :: Maybe Integer

A value of the long data type.

booleanValue :: Maybe Bool

A value of the Boolean data type.

blobValue :: Maybe Base64

A value of the BLOB data type.

isNull :: Maybe Bool

A value that indicates whether the data is NULL.

newField :: Field

Create a value of Field with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:doubleValue:Field', field_doubleValue - A value of the double data type.

$sel:stringValue:Field', field_stringValue - A value of the string data type.

$sel:longValue:Field', field_longValue - A value of the long data type.

$sel:booleanValue:Field', field_booleanValue - A value of the Boolean data type.

$sel:blobValue:Field', field_blobValue - A value of the BLOB data type.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:isNull:Field', field_isNull - A value that indicates whether the data is NULL.

field_doubleValue :: Lens' Field (Maybe Double)

A value of the double data type.

field_stringValue :: Lens' Field (Maybe Text)

A value of the string data type.

field_longValue :: Lens' Field (Maybe Integer)

A value of the long data type.

field_booleanValue :: Lens' Field (Maybe Bool)

A value of the Boolean data type.

field_blobValue :: Lens' Field (Maybe ByteString)

A value of the BLOB data type.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

field_isNull :: Lens' Field (Maybe Bool)

A value that indicates whether the data is NULL.

SqlParameter

data SqlParameter

A parameter used in a SQL statement.

See: newSqlParameter smart constructor.

Constructors

SqlParameter'

Fields

name :: Text

The name of the parameter.

value :: Text

The value of the parameter. Amazon Redshift implicitly converts to the proper data type. For more inforation, see Data types in the Amazon Redshift Database Developer Guide.

newSqlParameter

Arguments

:: Text
$sel:name:SqlParameter'

-> Text
$sel:value:SqlParameter'

-> SqlParameter

Create a value of SqlParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:SqlParameter', sqlParameter_name - The name of the parameter.

$sel:value:SqlParameter', sqlParameter_value - The value of the parameter. Amazon Redshift implicitly converts to the proper data type. For more inforation, see Data types in the Amazon Redshift Database Developer Guide.

sqlParameter_name :: Lens' SqlParameter Text

The name of the parameter.

sqlParameter_value :: Lens' SqlParameter Text

The value of the parameter. Amazon Redshift implicitly converts to the proper data type. For more inforation, see Data types in the Amazon Redshift Database Developer Guide.

StatementData

data StatementData

The SQL statement to run.

See: newStatementData smart constructor.

Constructors

StatementData'

Fields

status :: Maybe StatusString

The status of the SQL statement. An example is the that the SQL statement finished.

createdAt :: Maybe POSIX

The date and time (UTC) the statement was created.

queryParameters :: Maybe (NonEmpty SqlParameter)

The parameters used in a SQL statement.

queryStrings :: Maybe [Text]

One or more SQL statements. Each query string in the array corresponds to one of the queries in a batch query request.

queryString :: Maybe Text

The SQL statement.

statementName :: Maybe Text

The name of the SQL statement.

updatedAt :: Maybe POSIX

The date and time (UTC) that the statement metadata was last updated.

secretArn :: Maybe Text

The name or Amazon Resource Name (ARN) of the secret that enables access to the database.

isBatchStatement :: Maybe Bool

A value that indicates whether the statement is a batch query request.

id :: Text

The SQL statement identifier. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API.

newStatementData

Arguments

:: Text
$sel:id:StatementData'

-> StatementData

Create a value of StatementData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StatementData', statementData_status - The status of the SQL statement. An example is the that the SQL statement finished.

$sel:createdAt:StatementData', statementData_createdAt - The date and time (UTC) the statement was created.

$sel:queryParameters:StatementData', statementData_queryParameters - The parameters used in a SQL statement.

$sel:queryStrings:StatementData', statementData_queryStrings - One or more SQL statements. Each query string in the array corresponds to one of the queries in a batch query request.

$sel:queryString:StatementData', statementData_queryString - The SQL statement.

$sel:statementName:StatementData', statementData_statementName - The name of the SQL statement.

$sel:updatedAt:StatementData', statementData_updatedAt - The date and time (UTC) that the statement metadata was last updated.

$sel:secretArn:StatementData', statementData_secretArn - The name or Amazon Resource Name (ARN) of the secret that enables access to the database.

$sel:isBatchStatement:StatementData', statementData_isBatchStatement - A value that indicates whether the statement is a batch query request.

$sel:id:StatementData', statementData_id - The SQL statement identifier. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API.

statementData_status :: Lens' StatementData (Maybe StatusString)

The status of the SQL statement. An example is the that the SQL statement finished.

statementData_createdAt :: Lens' StatementData (Maybe UTCTime)

The date and time (UTC) the statement was created.

statementData_queryParameters :: Lens' StatementData (Maybe (NonEmpty SqlParameter))

The parameters used in a SQL statement.

statementData_queryStrings :: Lens' StatementData (Maybe [Text])

One or more SQL statements. Each query string in the array corresponds to one of the queries in a batch query request.

statementData_queryString :: Lens' StatementData (Maybe Text)

The SQL statement.

statementData_statementName :: Lens' StatementData (Maybe Text)

The name of the SQL statement.

statementData_updatedAt :: Lens' StatementData (Maybe UTCTime)

The date and time (UTC) that the statement metadata was last updated.

statementData_secretArn :: Lens' StatementData (Maybe Text)

The name or Amazon Resource Name (ARN) of the secret that enables access to the database.

statementData_isBatchStatement :: Lens' StatementData (Maybe Bool)

A value that indicates whether the statement is a batch query request.

statementData_id :: Lens' StatementData Text

The SQL statement identifier. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API.

SubStatementData

data SubStatementData

Information about an SQL statement.

See: newSubStatementData smart constructor.

Constructors

SubStatementData'

Fields

status :: Maybe StatementStatusString

The status of the SQL statement. An example is the that the SQL statement finished.

redshiftQueryId :: Maybe Integer

The SQL statement identifier. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API.

resultSize :: Maybe Integer

The size in bytes of the returned results. A -1 indicates the value is null.

createdAt :: Maybe POSIX

The date and time (UTC) the statement was created.

error :: Maybe Text

The error message from the cluster if the SQL statement encountered an error while running.

resultRows :: Maybe Integer

Either the number of rows returned from the SQL statement or the number of rows affected. If result size is greater than zero, the result rows can be the number of rows affected by SQL statements such as INSERT, UPDATE, DELETE, COPY, and others. A -1 indicates the value is null.

hasResultSet :: Maybe Bool

A value that indicates whether the statement has a result set. The result set can be empty.

queryString :: Maybe Text

The SQL statement text.

updatedAt :: Maybe POSIX

The date and time (UTC) that the statement metadata was last updated.

duration :: Maybe Integer

The amount of time in nanoseconds that the statement ran.

id :: Text

The identifier of the SQL statement. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API. A suffix indicates the number of the SQL statement. For example, d9b6c0c9-0747-4bf4-b142-e8883122f766:2 has a suffix of :2 that indicates the second SQL statement of a batch query.

newSubStatementData

Arguments

:: Text
$sel:id:SubStatementData'

-> SubStatementData

Create a value of SubStatementData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SubStatementData', subStatementData_status - The status of the SQL statement. An example is the that the SQL statement finished.

$sel:redshiftQueryId:SubStatementData', subStatementData_redshiftQueryId - The SQL statement identifier. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API.

$sel:resultSize:SubStatementData', subStatementData_resultSize - The size in bytes of the returned results. A -1 indicates the value is null.

$sel:createdAt:SubStatementData', subStatementData_createdAt - The date and time (UTC) the statement was created.

$sel:error:SubStatementData', subStatementData_error - The error message from the cluster if the SQL statement encountered an error while running.

$sel:resultRows:SubStatementData', subStatementData_resultRows - Either the number of rows returned from the SQL statement or the number of rows affected. If result size is greater than zero, the result rows can be the number of rows affected by SQL statements such as INSERT, UPDATE, DELETE, COPY, and others. A -1 indicates the value is null.

$sel:hasResultSet:SubStatementData', subStatementData_hasResultSet - A value that indicates whether the statement has a result set. The result set can be empty.

$sel:queryString:SubStatementData', subStatementData_queryString - The SQL statement text.

$sel:updatedAt:SubStatementData', subStatementData_updatedAt - The date and time (UTC) that the statement metadata was last updated.

$sel:duration:SubStatementData', subStatementData_duration - The amount of time in nanoseconds that the statement ran.

$sel:id:SubStatementData', subStatementData_id - The identifier of the SQL statement. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API. A suffix indicates the number of the SQL statement. For example, d9b6c0c9-0747-4bf4-b142-e8883122f766:2 has a suffix of :2 that indicates the second SQL statement of a batch query.

subStatementData_status :: Lens' SubStatementData (Maybe StatementStatusString)

The status of the SQL statement. An example is the that the SQL statement finished.

subStatementData_redshiftQueryId :: Lens' SubStatementData (Maybe Integer)

The SQL statement identifier. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API.

subStatementData_resultSize :: Lens' SubStatementData (Maybe Integer)

The size in bytes of the returned results. A -1 indicates the value is null.

subStatementData_createdAt :: Lens' SubStatementData (Maybe UTCTime)

The date and time (UTC) the statement was created.

subStatementData_error :: Lens' SubStatementData (Maybe Text)

The error message from the cluster if the SQL statement encountered an error while running.

subStatementData_resultRows :: Lens' SubStatementData (Maybe Integer)

Either the number of rows returned from the SQL statement or the number of rows affected. If result size is greater than zero, the result rows can be the number of rows affected by SQL statements such as INSERT, UPDATE, DELETE, COPY, and others. A -1 indicates the value is null.

subStatementData_hasResultSet :: Lens' SubStatementData (Maybe Bool)

A value that indicates whether the statement has a result set. The result set can be empty.

subStatementData_queryString :: Lens' SubStatementData (Maybe Text)

The SQL statement text.

subStatementData_updatedAt :: Lens' SubStatementData (Maybe UTCTime)

The date and time (UTC) that the statement metadata was last updated.

subStatementData_duration :: Lens' SubStatementData (Maybe Integer)

The amount of time in nanoseconds that the statement ran.

subStatementData_id :: Lens' SubStatementData Text

The identifier of the SQL statement. This value is a universally unique identifier (UUID) generated by Amazon Redshift Data API. A suffix indicates the number of the SQL statement. For example, d9b6c0c9-0747-4bf4-b142-e8883122f766:2 has a suffix of :2 that indicates the second SQL statement of a batch query.

TableMember

data TableMember

The properties of a table.

See: newTableMember smart constructor.

Constructors

TableMember'

Fields

schema :: Maybe Text

The schema containing the table.

name :: Maybe Text

The name of the table.

type' :: Maybe Text

The type of the table. Possible values include TABLE, VIEW, SYSTEM TABLE, GLOBAL TEMPORARY, LOCAL TEMPORARY, ALIAS, and SYNONYM.

newTableMember :: TableMember

Create a value of TableMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schema:TableMember', tableMember_schema - The schema containing the table.

$sel:name:TableMember', tableMember_name - The name of the table.

$sel:type':TableMember', tableMember_type - The type of the table. Possible values include TABLE, VIEW, SYSTEM TABLE, GLOBAL TEMPORARY, LOCAL TEMPORARY, ALIAS, and SYNONYM.

tableMember_schema :: Lens' TableMember (Maybe Text)

The schema containing the table.

tableMember_name :: Lens' TableMember (Maybe Text)

The name of the table.

tableMember_type :: Lens' TableMember (Maybe Text)

The type of the table. Possible values include TABLE, VIEW, SYSTEM TABLE, GLOBAL TEMPORARY, LOCAL TEMPORARY, ALIAS, and SYNONYM.