Amazonka.RedshiftData.Types.SqlParameter

Description

Documentation

data SqlParameter

A parameter used in a SQL statement.

See: newSqlParameter smart constructor.

Constructors

SqlParameter'

Fields

name :: Text

The name of the parameter.

value :: Text

The value of the parameter. Amazon Redshift implicitly converts to the proper data type. For more inforation, see Data types in the Amazon Redshift Database Developer Guide.

newSqlParameter

Arguments

:: Text
$sel:name:SqlParameter'

-> Text
$sel:value:SqlParameter'

-> SqlParameter

Create a value of SqlParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:SqlParameter', sqlParameter_name - The name of the parameter.

$sel:value:SqlParameter', sqlParameter_value - The value of the parameter. Amazon Redshift implicitly converts to the proper data type. For more inforation, see Data types in the Amazon Redshift Database Developer Guide.

sqlParameter_name :: Lens' SqlParameter Text

The name of the parameter.

sqlParameter_value :: Lens' SqlParameter Text

The value of the parameter. Amazon Redshift implicitly converts to the proper data type. For more inforation, see Data types in the Amazon Redshift Database Developer Guide.