Amazonka.Rekognition.CreateStreamProcessor

Description

Creates an Amazon Rekognition stream processor that you can use to detect and recognize faces in a streaming video.

Amazon Rekognition Video is a consumer of live video from Amazon Kinesis Video Streams. Amazon Rekognition Video sends analysis results to Amazon Kinesis Data Streams.

You provide as input a Kinesis video stream (Input) and a Kinesis data stream (Output) stream. You also specify the face recognition criteria in Settings. For example, the collection containing faces that you want to recognize. Use Name to assign an identifier for the stream processor. You use Name to manage the stream processor. For example, you can start processing the source video by calling StartStreamProcessor with the Name field.

After you have finished analyzing a streaming video, use StopStreamProcessor to stop processing. You can delete the stream processor by calling DeleteStreamProcessor.

This operation requires permissions to perform the rekognition:CreateStreamProcessor action. If you want to tag your stream processor, you also require permission to perform the rekognition:TagResource operation.

Creating a Request

data CreateStreamProcessor

See: newCreateStreamProcessor smart constructor.

Constructors

CreateStreamProcessor'

Fields

tags :: Maybe (HashMap Text Text)

A set of tags (key-value pairs) that you want to attach to the stream processor.

input :: StreamProcessorInput

Kinesis video stream stream that provides the source streaming video. If you are using the AWS CLI, the parameter name is StreamProcessorInput.

output :: StreamProcessorOutput

Kinesis data stream stream to which Amazon Rekognition Video puts the analysis results. If you are using the AWS CLI, the parameter name is StreamProcessorOutput.

name :: Text

An identifier you assign to the stream processor. You can use Name to manage the stream processor. For example, you can get the current status of the stream processor by calling DescribeStreamProcessor. Name is idempotent.

settings :: StreamProcessorSettings

Face recognition input parameters to be used by the stream processor. Includes the collection to use for face recognition and the face attributes to detect.

roleArn :: Text

ARN of the IAM role that allows access to the stream processor.

newCreateStreamProcessor

Arguments

:: StreamProcessorInput
$sel:input:CreateStreamProcessor'

-> StreamProcessorOutput
$sel:output:CreateStreamProcessor'

-> Text
$sel:name:CreateStreamProcessor'

-> StreamProcessorSettings
$sel:settings:CreateStreamProcessor'

-> Text
$sel:roleArn:CreateStreamProcessor'

-> CreateStreamProcessor

Create a value of CreateStreamProcessor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateStreamProcessor', createStreamProcessor_tags - A set of tags (key-value pairs) that you want to attach to the stream processor.

$sel:input:CreateStreamProcessor', createStreamProcessor_input - Kinesis video stream stream that provides the source streaming video. If you are using the AWS CLI, the parameter name is StreamProcessorInput.

$sel:output:CreateStreamProcessor', createStreamProcessor_output - Kinesis data stream stream to which Amazon Rekognition Video puts the analysis results. If you are using the AWS CLI, the parameter name is StreamProcessorOutput.

$sel:name:CreateStreamProcessor', createStreamProcessor_name - An identifier you assign to the stream processor. You can use Name to manage the stream processor. For example, you can get the current status of the stream processor by calling DescribeStreamProcessor. Name is idempotent.

$sel:settings:CreateStreamProcessor', createStreamProcessor_settings - Face recognition input parameters to be used by the stream processor. Includes the collection to use for face recognition and the face attributes to detect.

$sel:roleArn:CreateStreamProcessor', createStreamProcessor_roleArn - ARN of the IAM role that allows access to the stream processor.

Request Lenses

createStreamProcessor_tags :: Lens' CreateStreamProcessor (Maybe (HashMap Text Text))

A set of tags (key-value pairs) that you want to attach to the stream processor.

createStreamProcessor_input :: Lens' CreateStreamProcessor StreamProcessorInput

Kinesis video stream stream that provides the source streaming video. If you are using the AWS CLI, the parameter name is StreamProcessorInput.

createStreamProcessor_output :: Lens' CreateStreamProcessor StreamProcessorOutput

Kinesis data stream stream to which Amazon Rekognition Video puts the analysis results. If you are using the AWS CLI, the parameter name is StreamProcessorOutput.

createStreamProcessor_name :: Lens' CreateStreamProcessor Text

An identifier you assign to the stream processor. You can use Name to manage the stream processor. For example, you can get the current status of the stream processor by calling DescribeStreamProcessor. Name is idempotent.

createStreamProcessor_settings :: Lens' CreateStreamProcessor StreamProcessorSettings

Face recognition input parameters to be used by the stream processor. Includes the collection to use for face recognition and the face attributes to detect.

createStreamProcessor_roleArn :: Lens' CreateStreamProcessor Text

ARN of the IAM role that allows access to the stream processor.

Destructuring the Response

data CreateStreamProcessorResponse

See: newCreateStreamProcessorResponse smart constructor.

Constructors

CreateStreamProcessorResponse'

Fields

streamProcessorArn :: Maybe Text

ARN for the newly create stream processor.

httpStatus :: Int

The response's http status code.

newCreateStreamProcessorResponse

Arguments

:: Int
$sel:httpStatus:CreateStreamProcessorResponse'

-> CreateStreamProcessorResponse

Create a value of CreateStreamProcessorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamProcessorArn:CreateStreamProcessorResponse', createStreamProcessorResponse_streamProcessorArn - ARN for the newly create stream processor.

$sel:httpStatus:CreateStreamProcessorResponse', createStreamProcessorResponse_httpStatus - The response's http status code.

Response Lenses

createStreamProcessorResponse_streamProcessorArn :: Lens' CreateStreamProcessorResponse (Maybe Text)

ARN for the newly create stream processor.

createStreamProcessorResponse_httpStatus :: Lens' CreateStreamProcessorResponse Int

The response's http status code.