Amazonka.Rekognition.DescribeProjects

Description

Lists and gets information about your Amazon Rekognition Custom Labels projects.

This operation requires permissions to perform the rekognition:DescribeProjects action.

This operation returns paginated results.

Creating a Request

data DescribeProjects

See: newDescribeProjects smart constructor.

Constructors

DescribeProjects'

Fields

nextToken :: Maybe Text

If the previous response was incomplete (because there is more results to retrieve), Amazon Rekognition Custom Labels returns a pagination token in the response. You can use this pagination token to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per paginated call. The largest value you can specify is 100. If you specify a value greater than 100, a ValidationException error occurs. The default value is 100.

newDescribeProjects :: DescribeProjects

Create a value of DescribeProjects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeProjects', describeProjects_nextToken - If the previous response was incomplete (because there is more results to retrieve), Amazon Rekognition Custom Labels returns a pagination token in the response. You can use this pagination token to retrieve the next set of results.

$sel:maxResults:DescribeProjects', describeProjects_maxResults - The maximum number of results to return per paginated call. The largest value you can specify is 100. If you specify a value greater than 100, a ValidationException error occurs. The default value is 100.

Request Lenses

describeProjects_nextToken :: Lens' DescribeProjects (Maybe Text)

If the previous response was incomplete (because there is more results to retrieve), Amazon Rekognition Custom Labels returns a pagination token in the response. You can use this pagination token to retrieve the next set of results.

describeProjects_maxResults :: Lens' DescribeProjects (Maybe Natural)

The maximum number of results to return per paginated call. The largest value you can specify is 100. If you specify a value greater than 100, a ValidationException error occurs. The default value is 100.

Destructuring the Response

data DescribeProjectsResponse

See: newDescribeProjectsResponse smart constructor.

Constructors

DescribeProjectsResponse'

Fields

nextToken :: Maybe Text

If the previous response was incomplete (because there is more results to retrieve), Amazon Rekognition Custom Labels returns a pagination token in the response. You can use this pagination token to retrieve the next set of results.

projectDescriptions :: Maybe [ProjectDescription]

A list of project descriptions. The list is sorted by the date and time the projects are created.

httpStatus :: Int

The response's http status code.

newDescribeProjectsResponse

Arguments

:: Int
$sel:httpStatus:DescribeProjectsResponse'

-> DescribeProjectsResponse

Create a value of DescribeProjectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeProjects', describeProjectsResponse_nextToken - If the previous response was incomplete (because there is more results to retrieve), Amazon Rekognition Custom Labels returns a pagination token in the response. You can use this pagination token to retrieve the next set of results.

$sel:projectDescriptions:DescribeProjectsResponse', describeProjectsResponse_projectDescriptions - A list of project descriptions. The list is sorted by the date and time the projects are created.

$sel:httpStatus:DescribeProjectsResponse', describeProjectsResponse_httpStatus - The response's http status code.

Response Lenses

describeProjectsResponse_nextToken :: Lens' DescribeProjectsResponse (Maybe Text)

If the previous response was incomplete (because there is more results to retrieve), Amazon Rekognition Custom Labels returns a pagination token in the response. You can use this pagination token to retrieve the next set of results.

describeProjectsResponse_projectDescriptions :: Lens' DescribeProjectsResponse (Maybe [ProjectDescription])

A list of project descriptions. The list is sorted by the date and time the projects are created.

describeProjectsResponse_httpStatus :: Lens' DescribeProjectsResponse Int

The response's http status code.