Amazonka.Rekognition.SearchFacesByImage

Description

For a given input image, first detects the largest face in the image, and then searches the specified collection for matching faces. The operation compares the features of the input face with faces in the specified collection.

To search for all faces in an input image, you might first call the IndexFaces operation, and then use the face IDs returned in subsequent calls to the SearchFaces operation.

You can also call the DetectFaces operation and use the bounding boxes in the response to make face crops, which then you can pass in to the SearchFacesByImage operation.

You pass the input image either as base64-encoded image bytes or as a reference to an image in an Amazon S3 bucket. If you use the AWS CLI to call Amazon Rekognition operations, passing image bytes is not supported. The image must be either a PNG or JPEG formatted file.

The response returns an array of faces that match, ordered by similarity score with the highest similarity first. More specifically, it is an array of metadata for each face match found. Along with the metadata, the response also includes a similarity indicating how similar the face is to the input face. In the response, the operation also returns the bounding box (and a confidence level that the bounding box contains a face) of the face that Amazon Rekognition used for the input image.

If no faces are detected in the input image, SearchFacesByImage returns an InvalidParameterException error.

For an example, Searching for a Face Using an Image in the Amazon Rekognition Developer Guide.

The QualityFilter input parameter allows you to filter out detected faces that don’t meet a required quality bar. The quality bar is based on a variety of common use cases. Use QualityFilter to set the quality bar for filtering by specifying LOW, MEDIUM, or HIGH. If you do not want to filter detected faces, specify NONE. The default value is NONE.

To use quality filtering, you need a collection associated with version 3 of the face model or higher. To get the version of the face model associated with a collection, call DescribeCollection.

This operation requires permissions to perform the rekognition:SearchFacesByImage action.

Creating a Request

data SearchFacesByImage

See: newSearchFacesByImage smart constructor.

Constructors

SearchFacesByImage'

Fields

qualityFilter :: Maybe QualityFilter

A filter that specifies a quality bar for how much filtering is done to identify faces. Filtered faces aren't searched for in the collection. If you specify AUTO, Amazon Rekognition chooses the quality bar. If you specify LOW, MEDIUM, or HIGH, filtering removes all faces that don’t meet the chosen quality bar. The quality bar is based on a variety of common use cases. Low-quality detections can occur for a number of reasons. Some examples are an object that's misidentified as a face, a face that's too blurry, or a face with a pose that's too extreme to use. If you specify NONE, no filtering is performed. The default value is NONE.

To use quality filtering, the collection you are using must be associated with version 3 of the face model or higher.

faceMatchThreshold :: Maybe Double

(Optional) Specifies the minimum confidence in the face match to return. For example, don't return any matches where confidence in matches is less than 70%. The default value is 80%.

maxFaces :: Maybe Natural

Maximum number of faces to return. The operation returns the maximum number of faces with the highest confidence in the match.

collectionId :: Text

ID of the collection to search.

image :: Image

The input image as base64-encoded bytes or an S3 object. If you use the AWS CLI to call Amazon Rekognition operations, passing base64-encoded image bytes is not supported.

If you are using an AWS SDK to call Amazon Rekognition, you might not need to base64-encode image bytes passed using the Bytes field. For more information, see Images in the Amazon Rekognition developer guide.

newSearchFacesByImage

Arguments

:: Text
$sel:collectionId:SearchFacesByImage'

-> Image
$sel:image:SearchFacesByImage'

-> SearchFacesByImage

Create a value of SearchFacesByImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualityFilter:SearchFacesByImage', searchFacesByImage_qualityFilter - A filter that specifies a quality bar for how much filtering is done to identify faces. Filtered faces aren't searched for in the collection. If you specify AUTO, Amazon Rekognition chooses the quality bar. If you specify LOW, MEDIUM, or HIGH, filtering removes all faces that don’t meet the chosen quality bar. The quality bar is based on a variety of common use cases. Low-quality detections can occur for a number of reasons. Some examples are an object that's misidentified as a face, a face that's too blurry, or a face with a pose that's too extreme to use. If you specify NONE, no filtering is performed. The default value is NONE.

To use quality filtering, the collection you are using must be associated with version 3 of the face model or higher.

$sel:faceMatchThreshold:SearchFacesByImage', searchFacesByImage_faceMatchThreshold - (Optional) Specifies the minimum confidence in the face match to return. For example, don't return any matches where confidence in matches is less than 70%. The default value is 80%.

$sel:maxFaces:SearchFacesByImage', searchFacesByImage_maxFaces - Maximum number of faces to return. The operation returns the maximum number of faces with the highest confidence in the match.

$sel:collectionId:SearchFacesByImage', searchFacesByImage_collectionId - ID of the collection to search.

$sel:image:SearchFacesByImage', searchFacesByImage_image - The input image as base64-encoded bytes or an S3 object. If you use the AWS CLI to call Amazon Rekognition operations, passing base64-encoded image bytes is not supported.

If you are using an AWS SDK to call Amazon Rekognition, you might not need to base64-encode image bytes passed using the Bytes field. For more information, see Images in the Amazon Rekognition developer guide.

Request Lenses

searchFacesByImage_qualityFilter :: Lens' SearchFacesByImage (Maybe QualityFilter)

A filter that specifies a quality bar for how much filtering is done to identify faces. Filtered faces aren't searched for in the collection. If you specify AUTO, Amazon Rekognition chooses the quality bar. If you specify LOW, MEDIUM, or HIGH, filtering removes all faces that don’t meet the chosen quality bar. The quality bar is based on a variety of common use cases. Low-quality detections can occur for a number of reasons. Some examples are an object that's misidentified as a face, a face that's too blurry, or a face with a pose that's too extreme to use. If you specify NONE, no filtering is performed. The default value is NONE.

To use quality filtering, the collection you are using must be associated with version 3 of the face model or higher.

searchFacesByImage_faceMatchThreshold :: Lens' SearchFacesByImage (Maybe Double)

(Optional) Specifies the minimum confidence in the face match to return. For example, don't return any matches where confidence in matches is less than 70%. The default value is 80%.

searchFacesByImage_maxFaces :: Lens' SearchFacesByImage (Maybe Natural)

Maximum number of faces to return. The operation returns the maximum number of faces with the highest confidence in the match.

searchFacesByImage_collectionId :: Lens' SearchFacesByImage Text

ID of the collection to search.

searchFacesByImage_image :: Lens' SearchFacesByImage Image

The input image as base64-encoded bytes or an S3 object. If you use the AWS CLI to call Amazon Rekognition operations, passing base64-encoded image bytes is not supported.

If you are using an AWS SDK to call Amazon Rekognition, you might not need to base64-encode image bytes passed using the Bytes field. For more information, see Images in the Amazon Rekognition developer guide.

Destructuring the Response

data SearchFacesByImageResponse

See: newSearchFacesByImageResponse smart constructor.

Constructors

SearchFacesByImageResponse'

Fields

faceMatches :: Maybe [FaceMatch]

An array of faces that match the input face, along with the confidence in the match.

faceModelVersion :: Maybe Text

Version number of the face detection model associated with the input collection (CollectionId).

searchedFaceBoundingBox :: Maybe BoundingBox

The bounding box around the face in the input image that Amazon Rekognition used for the search.

searchedFaceConfidence :: Maybe Double

The level of confidence that the searchedFaceBoundingBox, contains a face.

httpStatus :: Int

The response's http status code.

newSearchFacesByImageResponse

Arguments

:: Int
$sel:httpStatus:SearchFacesByImageResponse'

-> SearchFacesByImageResponse

Create a value of SearchFacesByImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:faceMatches:SearchFacesByImageResponse', searchFacesByImageResponse_faceMatches - An array of faces that match the input face, along with the confidence in the match.

$sel:faceModelVersion:SearchFacesByImageResponse', searchFacesByImageResponse_faceModelVersion - Version number of the face detection model associated with the input collection (CollectionId).

$sel:searchedFaceBoundingBox:SearchFacesByImageResponse', searchFacesByImageResponse_searchedFaceBoundingBox - The bounding box around the face in the input image that Amazon Rekognition used for the search.

$sel:searchedFaceConfidence:SearchFacesByImageResponse', searchFacesByImageResponse_searchedFaceConfidence - The level of confidence that the searchedFaceBoundingBox, contains a face.

$sel:httpStatus:SearchFacesByImageResponse', searchFacesByImageResponse_httpStatus - The response's http status code.

Response Lenses

searchFacesByImageResponse_faceMatches :: Lens' SearchFacesByImageResponse (Maybe [FaceMatch])

An array of faces that match the input face, along with the confidence in the match.

searchFacesByImageResponse_faceModelVersion :: Lens' SearchFacesByImageResponse (Maybe Text)

Version number of the face detection model associated with the input collection (CollectionId).

searchFacesByImageResponse_searchedFaceBoundingBox :: Lens' SearchFacesByImageResponse (Maybe BoundingBox)

The bounding box around the face in the input image that Amazon Rekognition used for the search.

searchFacesByImageResponse_searchedFaceConfidence :: Lens' SearchFacesByImageResponse (Maybe Double)

The level of confidence that the searchedFaceBoundingBox, contains a face.

searchFacesByImageResponse_httpStatus :: Lens' SearchFacesByImageResponse Int

The response's http status code.