Amazonka.Rekognition.StartFaceDetection

Description

Starts asynchronous detection of faces in a stored video.

Amazon Rekognition Video can detect faces in a video stored in an Amazon S3 bucket. Use Video to specify the bucket name and the filename of the video. StartFaceDetection returns a job identifier (JobId) that you use to get the results of the operation. When face detection is finished, Amazon Rekognition Video publishes a completion status to the Amazon Simple Notification Service topic that you specify in NotificationChannel. To get the results of the face detection operation, first check that the status value published to the Amazon SNS topic is SUCCEEDED. If so, call GetFaceDetection and pass the job identifier (JobId) from the initial call to StartFaceDetection.

For more information, see Detecting Faces in a Stored Video in the Amazon Rekognition Developer Guide.

Creating a Request

data StartFaceDetection

See: newStartFaceDetection smart constructor.

Constructors

StartFaceDetection'

Fields

jobTag :: Maybe Text

An identifier you specify that's returned in the completion notification that's published to your Amazon Simple Notification Service topic. For example, you can use JobTag to group related jobs and identify them in the completion notification.

notificationChannel :: Maybe NotificationChannel

The ARN of the Amazon SNS topic to which you want Amazon Rekognition Video to publish the completion status of the face detection operation. The Amazon SNS topic must have a topic name that begins with AmazonRekognition if you are using the AmazonRekognitionServiceRole permissions policy.

clientRequestToken :: Maybe Text

Idempotent token used to identify the start request. If you use the same token with multiple StartFaceDetection requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidently started more than once.

faceAttributes :: Maybe FaceAttributes

The face attributes you want returned.

DEFAULT - The following subset of facial attributes are returned: BoundingBox, Confidence, Pose, Quality and Landmarks.

ALL - All facial attributes are returned.

video :: Video

The video in which you want to detect faces. The video must be stored in an Amazon S3 bucket.

newStartFaceDetection

Arguments

:: Video
$sel:video:StartFaceDetection'

-> StartFaceDetection

Create a value of StartFaceDetection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobTag:StartFaceDetection', startFaceDetection_jobTag - An identifier you specify that's returned in the completion notification that's published to your Amazon Simple Notification Service topic. For example, you can use JobTag to group related jobs and identify them in the completion notification.

$sel:notificationChannel:StartFaceDetection', startFaceDetection_notificationChannel - The ARN of the Amazon SNS topic to which you want Amazon Rekognition Video to publish the completion status of the face detection operation. The Amazon SNS topic must have a topic name that begins with AmazonRekognition if you are using the AmazonRekognitionServiceRole permissions policy.

$sel:clientRequestToken:StartFaceDetection', startFaceDetection_clientRequestToken - Idempotent token used to identify the start request. If you use the same token with multiple StartFaceDetection requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidently started more than once.

$sel:faceAttributes:StartFaceDetection', startFaceDetection_faceAttributes - The face attributes you want returned.

DEFAULT - The following subset of facial attributes are returned: BoundingBox, Confidence, Pose, Quality and Landmarks.

ALL - All facial attributes are returned.

$sel:video:StartFaceDetection', startFaceDetection_video - The video in which you want to detect faces. The video must be stored in an Amazon S3 bucket.

Request Lenses

startFaceDetection_jobTag :: Lens' StartFaceDetection (Maybe Text)

An identifier you specify that's returned in the completion notification that's published to your Amazon Simple Notification Service topic. For example, you can use JobTag to group related jobs and identify them in the completion notification.

startFaceDetection_notificationChannel :: Lens' StartFaceDetection (Maybe NotificationChannel)

The ARN of the Amazon SNS topic to which you want Amazon Rekognition Video to publish the completion status of the face detection operation. The Amazon SNS topic must have a topic name that begins with AmazonRekognition if you are using the AmazonRekognitionServiceRole permissions policy.

startFaceDetection_clientRequestToken :: Lens' StartFaceDetection (Maybe Text)

Idempotent token used to identify the start request. If you use the same token with multiple StartFaceDetection requests, the same JobId is returned. Use ClientRequestToken to prevent the same job from being accidently started more than once.

startFaceDetection_faceAttributes :: Lens' StartFaceDetection (Maybe FaceAttributes)

The face attributes you want returned.

DEFAULT - The following subset of facial attributes are returned: BoundingBox, Confidence, Pose, Quality and Landmarks.

ALL - All facial attributes are returned.

startFaceDetection_video :: Lens' StartFaceDetection Video

The video in which you want to detect faces. The video must be stored in an Amazon S3 bucket.

Destructuring the Response

data StartFaceDetectionResponse

See: newStartFaceDetectionResponse smart constructor.

Constructors

StartFaceDetectionResponse'

Fields

jobId :: Maybe Text

The identifier for the face detection job. Use JobId to identify the job in a subsequent call to GetFaceDetection.

httpStatus :: Int

The response's http status code.

newStartFaceDetectionResponse

Arguments

:: Int
$sel:httpStatus:StartFaceDetectionResponse'

-> StartFaceDetectionResponse

Create a value of StartFaceDetectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StartFaceDetectionResponse', startFaceDetectionResponse_jobId - The identifier for the face detection job. Use JobId to identify the job in a subsequent call to GetFaceDetection.

$sel:httpStatus:StartFaceDetectionResponse', startFaceDetectionResponse_httpStatus - The response's http status code.

Response Lenses

startFaceDetectionResponse_jobId :: Lens' StartFaceDetectionResponse (Maybe Text)

The identifier for the face detection job. Use JobId to identify the job in a subsequent call to GetFaceDetection.

startFaceDetectionResponse_httpStatus :: Lens' StartFaceDetectionResponse Int

The response's http status code.