Amazonka.Rekognition.StartStreamProcessor

Description

Starts processing a stream processor. You create a stream processor by calling CreateStreamProcessor. To tell StartStreamProcessor which stream processor to start, use the value of the Name field specified in the call to CreateStreamProcessor.

Creating a Request

data StartStreamProcessor

See: newStartStreamProcessor smart constructor.

Constructors

StartStreamProcessor'

Fields

name :: Text

The name of the stream processor to start processing.

newStartStreamProcessor

Arguments

:: Text
$sel:name:StartStreamProcessor'

-> StartStreamProcessor

Create a value of StartStreamProcessor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:StartStreamProcessor', startStreamProcessor_name - The name of the stream processor to start processing.

Request Lenses

startStreamProcessor_name :: Lens' StartStreamProcessor Text

The name of the stream processor to start processing.

Destructuring the Response

data StartStreamProcessorResponse

See: newStartStreamProcessorResponse smart constructor.

Constructors

StartStreamProcessorResponse'

Fields

httpStatus :: Int

The response's http status code.

newStartStreamProcessorResponse

Arguments

:: Int
$sel:httpStatus:StartStreamProcessorResponse'

-> StartStreamProcessorResponse

Create a value of StartStreamProcessorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartStreamProcessorResponse', startStreamProcessorResponse_httpStatus - The response's http status code.

Response Lenses

startStreamProcessorResponse_httpStatus :: Lens' StartStreamProcessorResponse Int

The response's http status code.