Amazonka.Rekognition.Types.AgeRange

Description

Documentation

data AgeRange

Structure containing the estimated age range, in years, for a face.

Amazon Rekognition estimates an age range for faces detected in the input image. Estimated age ranges can overlap. A face of a 5-year-old might have an estimated range of 4-6, while the face of a 6-year-old might have an estimated range of 4-8.

See: newAgeRange smart constructor.

Constructors

AgeRange'

Fields

low :: Maybe Natural

The lowest estimated age.

high :: Maybe Natural

The highest estimated age.

newAgeRange :: AgeRange

Create a value of AgeRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:low:AgeRange', ageRange_low - The lowest estimated age.

$sel:high:AgeRange', ageRange_high - The highest estimated age.

ageRange_low :: Lens' AgeRange (Maybe Natural)

The lowest estimated age.

ageRange_high :: Lens' AgeRange (Maybe Natural)

The highest estimated age.