Amazonka.Rekognition.Types.FaceDetail

Description

Documentation

data FaceDetail

Structure containing attributes of the face that the algorithm detected.

A FaceDetail object contains either the default facial attributes or all facial attributes. The default attributes are BoundingBox, Confidence, Landmarks, Pose, and Quality.

GetFaceDetection is the only Amazon Rekognition Video stored video operation that can return a FaceDetail object with all attributes. To specify which attributes to return, use the FaceAttributes input parameter for StartFaceDetection. The following Amazon Rekognition Video operations return only the default attributes. The corresponding Start operations don't have a FaceAttributes input parameter.

GetCelebrityRecognition

GetPersonTracking

GetFaceSearch

The Amazon Rekognition Image DetectFaces and IndexFaces operations can return all facial attributes. To specify which attributes to return, use the Attributes input parameter for DetectFaces. For IndexFaces, use the DetectAttributes input parameter.

See: newFaceDetail smart constructor.

Constructors

FaceDetail'

Fields

ageRange :: Maybe AgeRange

The estimated age range, in years, for the face. Low represents the lowest estimated age and High represents the highest estimated age.

sunglasses :: Maybe Sunglasses

Indicates whether or not the face is wearing sunglasses, and the confidence level in the determination.

mouthOpen :: Maybe MouthOpen

Indicates whether or not the mouth on the face is open, and the confidence level in the determination.

boundingBox :: Maybe BoundingBox

Bounding box of the face. Default attribute.

emotions :: Maybe [Emotion]

The emotions that appear to be expressed on the face, and the confidence level in the determination. The API is only making a determination of the physical appearance of a person's face. It is not a determination of the person’s internal emotional state and should not be used in such a way. For example, a person pretending to have a sad face might not be sad emotionally.

eyesOpen :: Maybe EyeOpen

Indicates whether or not the eyes on the face are open, and the confidence level in the determination.

pose :: Maybe Pose

Indicates the pose of the face as determined by its pitch, roll, and yaw. Default attribute.

confidence :: Maybe Double

Confidence level that the bounding box contains a face (and not a different object such as a tree). Default attribute.

gender :: Maybe Gender

The predicted gender of a detected face.

quality :: Maybe ImageQuality

Identifies image brightness and sharpness. Default attribute.

eyeglasses :: Maybe Eyeglasses

Indicates whether or not the face is wearing eye glasses, and the confidence level in the determination.

beard :: Maybe Beard

Indicates whether or not the face has a beard, and the confidence level in the determination.

mustache :: Maybe Mustache

Indicates whether or not the face has a mustache, and the confidence level in the determination.

smile :: Maybe Smile

Indicates whether or not the face is smiling, and the confidence level in the determination.

landmarks :: Maybe [Landmark]

Indicates the location of landmarks on the face. Default attribute.

newFaceDetail :: FaceDetail

Create a value of FaceDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ageRange:FaceDetail', faceDetail_ageRange - The estimated age range, in years, for the face. Low represents the lowest estimated age and High represents the highest estimated age.

$sel:sunglasses:FaceDetail', faceDetail_sunglasses - Indicates whether or not the face is wearing sunglasses, and the confidence level in the determination.

$sel:mouthOpen:FaceDetail', faceDetail_mouthOpen - Indicates whether or not the mouth on the face is open, and the confidence level in the determination.

$sel:boundingBox:FaceDetail', faceDetail_boundingBox - Bounding box of the face. Default attribute.

$sel:emotions:FaceDetail', faceDetail_emotions - The emotions that appear to be expressed on the face, and the confidence level in the determination. The API is only making a determination of the physical appearance of a person's face. It is not a determination of the person’s internal emotional state and should not be used in such a way. For example, a person pretending to have a sad face might not be sad emotionally.

$sel:eyesOpen:FaceDetail', faceDetail_eyesOpen - Indicates whether or not the eyes on the face are open, and the confidence level in the determination.

$sel:pose:FaceDetail', faceDetail_pose - Indicates the pose of the face as determined by its pitch, roll, and yaw. Default attribute.

$sel:confidence:FaceDetail', faceDetail_confidence - Confidence level that the bounding box contains a face (and not a different object such as a tree). Default attribute.

$sel:gender:FaceDetail', faceDetail_gender - The predicted gender of a detected face.

$sel:quality:FaceDetail', faceDetail_quality - Identifies image brightness and sharpness. Default attribute.

$sel:eyeglasses:FaceDetail', faceDetail_eyeglasses - Indicates whether or not the face is wearing eye glasses, and the confidence level in the determination.

$sel:beard:FaceDetail', faceDetail_beard - Indicates whether or not the face has a beard, and the confidence level in the determination.

$sel:mustache:FaceDetail', faceDetail_mustache - Indicates whether or not the face has a mustache, and the confidence level in the determination.

$sel:smile:FaceDetail', faceDetail_smile - Indicates whether or not the face is smiling, and the confidence level in the determination.

$sel:landmarks:FaceDetail', faceDetail_landmarks - Indicates the location of landmarks on the face. Default attribute.

faceDetail_ageRange :: Lens' FaceDetail (Maybe AgeRange)

The estimated age range, in years, for the face. Low represents the lowest estimated age and High represents the highest estimated age.

faceDetail_sunglasses :: Lens' FaceDetail (Maybe Sunglasses)

Indicates whether or not the face is wearing sunglasses, and the confidence level in the determination.

faceDetail_mouthOpen :: Lens' FaceDetail (Maybe MouthOpen)

Indicates whether or not the mouth on the face is open, and the confidence level in the determination.

faceDetail_boundingBox :: Lens' FaceDetail (Maybe BoundingBox)

Bounding box of the face. Default attribute.

faceDetail_emotions :: Lens' FaceDetail (Maybe [Emotion])

The emotions that appear to be expressed on the face, and the confidence level in the determination. The API is only making a determination of the physical appearance of a person's face. It is not a determination of the person’s internal emotional state and should not be used in such a way. For example, a person pretending to have a sad face might not be sad emotionally.

faceDetail_eyesOpen :: Lens' FaceDetail (Maybe EyeOpen)

Indicates whether or not the eyes on the face are open, and the confidence level in the determination.

faceDetail_pose :: Lens' FaceDetail (Maybe Pose)

Indicates the pose of the face as determined by its pitch, roll, and yaw. Default attribute.

faceDetail_confidence :: Lens' FaceDetail (Maybe Double)

Confidence level that the bounding box contains a face (and not a different object such as a tree). Default attribute.

faceDetail_gender :: Lens' FaceDetail (Maybe Gender)

The predicted gender of a detected face.

faceDetail_quality :: Lens' FaceDetail (Maybe ImageQuality)

Identifies image brightness and sharpness. Default attribute.

faceDetail_eyeglasses :: Lens' FaceDetail (Maybe Eyeglasses)

Indicates whether or not the face is wearing eye glasses, and the confidence level in the determination.

faceDetail_beard :: Lens' FaceDetail (Maybe Beard)

Indicates whether or not the face has a beard, and the confidence level in the determination.

faceDetail_mustache :: Lens' FaceDetail (Maybe Mustache)

Indicates whether or not the face has a mustache, and the confidence level in the determination.

faceDetail_smile :: Lens' FaceDetail (Maybe Smile)

Indicates whether or not the face is smiling, and the confidence level in the determination.

faceDetail_landmarks :: Lens' FaceDetail (Maybe [Landmark])

Indicates the location of landmarks on the face. Default attribute.