Amazonka.Rekognition.Types.Landmark

Description

Documentation

data Landmark

Indicates the location of the landmark on the face.

See: newLandmark smart constructor.

Constructors

Landmark'

Fields

type' :: Maybe LandmarkType

Type of landmark.

x :: Maybe Double

The x-coordinate of the landmark expressed as a ratio of the width of the image. The x-coordinate is measured from the left-side of the image. For example, if the image is 700 pixels wide and the x-coordinate of the landmark is at 350 pixels, this value is 0.5.

y :: Maybe Double

The y-coordinate of the landmark expressed as a ratio of the height of the image. The y-coordinate is measured from the top of the image. For example, if the image height is 200 pixels and the y-coordinate of the landmark is at 50 pixels, this value is 0.25.

newLandmark :: Landmark

Create a value of Landmark with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':Landmark', landmark_type - Type of landmark.

$sel:x:Landmark', landmark_x - The x-coordinate of the landmark expressed as a ratio of the width of the image. The x-coordinate is measured from the left-side of the image. For example, if the image is 700 pixels wide and the x-coordinate of the landmark is at 350 pixels, this value is 0.5.

$sel:y:Landmark', landmark_y - The y-coordinate of the landmark expressed as a ratio of the height of the image. The y-coordinate is measured from the top of the image. For example, if the image height is 200 pixels and the y-coordinate of the landmark is at 50 pixels, this value is 0.25.

landmark_type :: Lens' Landmark (Maybe LandmarkType)

Type of landmark.

landmark_x :: Lens' Landmark (Maybe Double)

The x-coordinate of the landmark expressed as a ratio of the width of the image. The x-coordinate is measured from the left-side of the image. For example, if the image is 700 pixels wide and the x-coordinate of the landmark is at 350 pixels, this value is 0.5.

landmark_y :: Lens' Landmark (Maybe Double)

The y-coordinate of the landmark expressed as a ratio of the height of the image. The y-coordinate is measured from the top of the image. For example, if the image height is 200 pixels and the y-coordinate of the landmark is at 50 pixels, this value is 0.25.