Amazonka.Rekognition.Types.Point

Description

Documentation

data Point

The X and Y coordinates of a point on an image. The X and Y values returned are ratios of the overall image size. For example, if the input image is 700x200 and the operation returns X=0.5 and Y=0.25, then the point is at the (350,50) pixel coordinate on the image.

An array of Point objects, Polygon, is returned by DetectText and by DetectCustomLabels. Polygon represents a fine-grained polygon around a detected item. For more information, see Geometry in the Amazon Rekognition Developer Guide.

See: newPoint smart constructor.

Constructors

Point'

Fields

x :: Maybe Double

The value of the X coordinate for a point on a Polygon.

y :: Maybe Double

The value of the Y coordinate for a point on a Polygon.

newPoint :: Point

Create a value of Point with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:x:Point', point_x - The value of the X coordinate for a point on a Polygon.

$sel:y:Point', point_y - The value of the Y coordinate for a point on a Polygon.

point_x :: Lens' Point (Maybe Double)

The value of the X coordinate for a point on a Polygon.

point_y :: Lens' Point (Maybe Double)

The value of the Y coordinate for a point on a Polygon.