Amazonka.Rekognition.Types.ProjectVersionDescription

Description

Documentation

data ProjectVersionDescription

The description of a version of a model.

See: newProjectVersionDescription smart constructor.

Constructors

ProjectVersionDescription'

Fields

minInferenceUnits :: Maybe Natural

The minimum number of inference units used by the model. For more information, see StartProjectVersion.

status :: Maybe ProjectVersionStatus

The current status of the model version.

evaluationResult :: Maybe EvaluationResult

The training results. EvaluationResult is only returned if training is successful.

manifestSummary :: Maybe GroundTruthManifest

The location of the summary manifest. The summary manifest provides aggregate data validation results for the training and test datasets.

kmsKeyId :: Maybe Text

The identifer for the AWS Key Management Service (AWS KMS) customer master key that was used to encrypt the model during training.

testingDataResult :: Maybe TestingDataResult

Contains information about the testing results.

statusMessage :: Maybe Text

A descriptive message for an error or warning that occurred.

creationTimestamp :: Maybe POSIX

The Unix datetime for the date and time that training started.

projectVersionArn :: Maybe Text

The Amazon Resource Name (ARN) of the model version.

outputConfig :: Maybe OutputConfig

The location where training results are saved.

billableTrainingTimeInSeconds :: Maybe Natural

The duration, in seconds, that the model version has been billed for training. This value is only returned if the model version has been successfully trained.

trainingEndTimestamp :: Maybe POSIX

The Unix date and time that training of the model ended.

trainingDataResult :: Maybe TrainingDataResult

Contains information about the training results.

newProjectVersionDescription :: ProjectVersionDescription

Create a value of ProjectVersionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minInferenceUnits:ProjectVersionDescription', projectVersionDescription_minInferenceUnits - The minimum number of inference units used by the model. For more information, see StartProjectVersion.

$sel:status:ProjectVersionDescription', projectVersionDescription_status - The current status of the model version.

$sel:evaluationResult:ProjectVersionDescription', projectVersionDescription_evaluationResult - The training results. EvaluationResult is only returned if training is successful.

$sel:manifestSummary:ProjectVersionDescription', projectVersionDescription_manifestSummary - The location of the summary manifest. The summary manifest provides aggregate data validation results for the training and test datasets.

$sel:kmsKeyId:ProjectVersionDescription', projectVersionDescription_kmsKeyId - The identifer for the AWS Key Management Service (AWS KMS) customer master key that was used to encrypt the model during training.

$sel:testingDataResult:ProjectVersionDescription', projectVersionDescription_testingDataResult - Contains information about the testing results.

$sel:statusMessage:ProjectVersionDescription', projectVersionDescription_statusMessage - A descriptive message for an error or warning that occurred.

$sel:creationTimestamp:ProjectVersionDescription', projectVersionDescription_creationTimestamp - The Unix datetime for the date and time that training started.

$sel:projectVersionArn:ProjectVersionDescription', projectVersionDescription_projectVersionArn - The Amazon Resource Name (ARN) of the model version.

$sel:outputConfig:ProjectVersionDescription', projectVersionDescription_outputConfig - The location where training results are saved.

$sel:billableTrainingTimeInSeconds:ProjectVersionDescription', projectVersionDescription_billableTrainingTimeInSeconds - The duration, in seconds, that the model version has been billed for training. This value is only returned if the model version has been successfully trained.

$sel:trainingEndTimestamp:ProjectVersionDescription', projectVersionDescription_trainingEndTimestamp - The Unix date and time that training of the model ended.

$sel:trainingDataResult:ProjectVersionDescription', projectVersionDescription_trainingDataResult - Contains information about the training results.

projectVersionDescription_minInferenceUnits :: Lens' ProjectVersionDescription (Maybe Natural)

The minimum number of inference units used by the model. For more information, see StartProjectVersion.

projectVersionDescription_status :: Lens' ProjectVersionDescription (Maybe ProjectVersionStatus)

The current status of the model version.

projectVersionDescription_evaluationResult :: Lens' ProjectVersionDescription (Maybe EvaluationResult)

The training results. EvaluationResult is only returned if training is successful.

projectVersionDescription_manifestSummary :: Lens' ProjectVersionDescription (Maybe GroundTruthManifest)

The location of the summary manifest. The summary manifest provides aggregate data validation results for the training and test datasets.

projectVersionDescription_kmsKeyId :: Lens' ProjectVersionDescription (Maybe Text)

The identifer for the AWS Key Management Service (AWS KMS) customer master key that was used to encrypt the model during training.

projectVersionDescription_testingDataResult :: Lens' ProjectVersionDescription (Maybe TestingDataResult)

Contains information about the testing results.

projectVersionDescription_statusMessage :: Lens' ProjectVersionDescription (Maybe Text)

A descriptive message for an error or warning that occurred.

projectVersionDescription_creationTimestamp :: Lens' ProjectVersionDescription (Maybe UTCTime)

The Unix datetime for the date and time that training started.

projectVersionDescription_projectVersionArn :: Lens' ProjectVersionDescription (Maybe Text)

The Amazon Resource Name (ARN) of the model version.

projectVersionDescription_outputConfig :: Lens' ProjectVersionDescription (Maybe OutputConfig)

The location where training results are saved.

projectVersionDescription_billableTrainingTimeInSeconds :: Lens' ProjectVersionDescription (Maybe Natural)

The duration, in seconds, that the model version has been billed for training. This value is only returned if the model version has been successfully trained.

projectVersionDescription_trainingEndTimestamp :: Lens' ProjectVersionDescription (Maybe UTCTime)

The Unix date and time that training of the model ended.

projectVersionDescription_trainingDataResult :: Lens' ProjectVersionDescription (Maybe TrainingDataResult)

Contains information about the training results.