Amazonka.Rekognition.Types.SegmentDetection

Description

Documentation

data SegmentDetection

A technical cue or shot detection segment detected in a video. An array of SegmentDetection objects containing all segments detected in a stored video is returned by GetSegmentDetection.

See: newSegmentDetection smart constructor.

Constructors

SegmentDetection'

Fields

technicalCueSegment :: Maybe TechnicalCueSegment

If the segment is a technical cue, contains information about the technical cue.

endFrameNumber :: Maybe Natural

The frame number at the end of a video segment, using a frame index that starts with 0.

durationSMPTE :: Maybe Text

The duration of the timecode for the detected segment in SMPTE format.

endTimestampMillis :: Maybe Integer

The end time of the detected segment, in milliseconds, from the start of the video. This value is rounded down.

startTimecodeSMPTE :: Maybe Text

The frame-accurate SMPTE timecode, from the start of a video, for the start of a detected segment. StartTimecode is in HH:MM:SS:fr format (and ;fr for drop frame-rates).

endTimecodeSMPTE :: Maybe Text

The frame-accurate SMPTE timecode, from the start of a video, for the end of a detected segment. EndTimecode is in HH:MM:SS:fr format (and ;fr for drop frame-rates).

durationMillis :: Maybe Natural

The duration of the detected segment in milliseconds.

durationFrames :: Maybe Natural

The duration of a video segment, expressed in frames.

startTimestampMillis :: Maybe Integer

The start time of the detected segment in milliseconds from the start of the video. This value is rounded down. For example, if the actual timestamp is 100.6667 milliseconds, Amazon Rekognition Video returns a value of 100 millis.

type' :: Maybe SegmentType

The type of the segment. Valid values are TECHNICAL_CUE and SHOT.

shotSegment :: Maybe ShotSegment

If the segment is a shot detection, contains information about the shot detection.

startFrameNumber :: Maybe Natural

The frame number of the start of a video segment, using a frame index that starts with 0.

newSegmentDetection :: SegmentDetection

Create a value of SegmentDetection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:technicalCueSegment:SegmentDetection', segmentDetection_technicalCueSegment - If the segment is a technical cue, contains information about the technical cue.

$sel:endFrameNumber:SegmentDetection', segmentDetection_endFrameNumber - The frame number at the end of a video segment, using a frame index that starts with 0.

$sel:durationSMPTE:SegmentDetection', segmentDetection_durationSMPTE - The duration of the timecode for the detected segment in SMPTE format.

$sel:endTimestampMillis:SegmentDetection', segmentDetection_endTimestampMillis - The end time of the detected segment, in milliseconds, from the start of the video. This value is rounded down.

$sel:startTimecodeSMPTE:SegmentDetection', segmentDetection_startTimecodeSMPTE - The frame-accurate SMPTE timecode, from the start of a video, for the start of a detected segment. StartTimecode is in HH:MM:SS:fr format (and ;fr for drop frame-rates).

$sel:endTimecodeSMPTE:SegmentDetection', segmentDetection_endTimecodeSMPTE - The frame-accurate SMPTE timecode, from the start of a video, for the end of a detected segment. EndTimecode is in HH:MM:SS:fr format (and ;fr for drop frame-rates).

$sel:durationMillis:SegmentDetection', segmentDetection_durationMillis - The duration of the detected segment in milliseconds.

$sel:durationFrames:SegmentDetection', segmentDetection_durationFrames - The duration of a video segment, expressed in frames.

$sel:startTimestampMillis:SegmentDetection', segmentDetection_startTimestampMillis - The start time of the detected segment in milliseconds from the start of the video. This value is rounded down. For example, if the actual timestamp is 100.6667 milliseconds, Amazon Rekognition Video returns a value of 100 millis.

$sel:type':SegmentDetection', segmentDetection_type - The type of the segment. Valid values are TECHNICAL_CUE and SHOT.

$sel:shotSegment:SegmentDetection', segmentDetection_shotSegment - If the segment is a shot detection, contains information about the shot detection.

$sel:startFrameNumber:SegmentDetection', segmentDetection_startFrameNumber - The frame number of the start of a video segment, using a frame index that starts with 0.

segmentDetection_technicalCueSegment :: Lens' SegmentDetection (Maybe TechnicalCueSegment)

If the segment is a technical cue, contains information about the technical cue.

segmentDetection_endFrameNumber :: Lens' SegmentDetection (Maybe Natural)

The frame number at the end of a video segment, using a frame index that starts with 0.

segmentDetection_durationSMPTE :: Lens' SegmentDetection (Maybe Text)

The duration of the timecode for the detected segment in SMPTE format.

segmentDetection_endTimestampMillis :: Lens' SegmentDetection (Maybe Integer)

The end time of the detected segment, in milliseconds, from the start of the video. This value is rounded down.

segmentDetection_startTimecodeSMPTE :: Lens' SegmentDetection (Maybe Text)

The frame-accurate SMPTE timecode, from the start of a video, for the start of a detected segment. StartTimecode is in HH:MM:SS:fr format (and ;fr for drop frame-rates).

segmentDetection_endTimecodeSMPTE :: Lens' SegmentDetection (Maybe Text)

The frame-accurate SMPTE timecode, from the start of a video, for the end of a detected segment. EndTimecode is in HH:MM:SS:fr format (and ;fr for drop frame-rates).

segmentDetection_durationMillis :: Lens' SegmentDetection (Maybe Natural)

The duration of the detected segment in milliseconds.

segmentDetection_durationFrames :: Lens' SegmentDetection (Maybe Natural)

The duration of a video segment, expressed in frames.

segmentDetection_startTimestampMillis :: Lens' SegmentDetection (Maybe Integer)

The start time of the detected segment in milliseconds from the start of the video. This value is rounded down. For example, if the actual timestamp is 100.6667 milliseconds, Amazon Rekognition Video returns a value of 100 millis.

segmentDetection_type :: Lens' SegmentDetection (Maybe SegmentType)

The type of the segment. Valid values are TECHNICAL_CUE and SHOT.

segmentDetection_shotSegment :: Lens' SegmentDetection (Maybe ShotSegment)

If the segment is a shot detection, contains information about the shot detection.

segmentDetection_startFrameNumber :: Lens' SegmentDetection (Maybe Natural)

The frame number of the start of a video segment, using a frame index that starts with 0.