Amazonka.ResourceGroups.Types.GroupFilter

Description

Documentation

data GroupFilter

A filter collection that you can use to restrict the results from a List operation to only those you want to include.

See: newGroupFilter smart constructor.

Constructors

GroupFilter'

Fields

name :: GroupFilterName

The name of the filter. Filter names are case-sensitive.

values :: NonEmpty Text

One or more filter values. Allowed filter values vary by group filter name, and are case-sensitive.

newGroupFilter

Arguments

:: GroupFilterName
$sel:name:GroupFilter'

-> NonEmpty Text
$sel:values:GroupFilter'

-> GroupFilter

Create a value of GroupFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GroupFilter', groupFilter_name - The name of the filter. Filter names are case-sensitive.

$sel:values:GroupFilter', groupFilter_values - One or more filter values. Allowed filter values vary by group filter name, and are case-sensitive.

groupFilter_name :: Lens' GroupFilter GroupFilterName

The name of the filter. Filter names are case-sensitive.

groupFilter_values :: Lens' GroupFilter (NonEmpty Text)

One or more filter values. Allowed filter values vary by group filter name, and are case-sensitive.