Amazonka.ResourceGroupsTagging.TagResources

Description

Applies one or more tags to the specified resources. Note the following:

Not all resources can have tags. For a list of services with resources that support tagging using this operation, see Services that support the Resource Groups Tagging API.

Each resource can have up to 50 tags. For other limits, see Tag Naming and Usage Conventions in the AWS General Reference.

You can only tag resources that are located in the specified AWS Region for the AWS account.

To add tags to a resource, you need the necessary permissions for the service that the resource belongs to as well as permissions for adding tags. For more information, see the documentation for each service.

Do not store personally identifiable information (PII) or other confidential or sensitive information in tags. We use tags to provide you with billing and administration services. Tags are not intended to be used for private or sensitive data.

Creating a Request

data TagResources

See: newTagResources smart constructor.

Constructors

TagResources'

Fields

resourceARNList :: NonEmpty Text

Specifies the list of ARNs of the resources that you want to apply tags to.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

tags :: HashMap Text Text

Specifies a list of tags that you want to add to the specified resources. A tag consists of a key and a value that you define.

newTagResources

Arguments

:: NonEmpty Text
$sel:resourceARNList:TagResources'

-> TagResources

Create a value of TagResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARNList:TagResources', tagResources_resourceARNList - Specifies the list of ARNs of the resources that you want to apply tags to.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

$sel:tags:TagResources', tagResources_tags - Specifies a list of tags that you want to add to the specified resources. A tag consists of a key and a value that you define.

Request Lenses

tagResources_resourceARNList :: Lens' TagResources (NonEmpty Text)

Specifies the list of ARNs of the resources that you want to apply tags to.

An ARN (Amazon Resource Name) uniquely identifies a resource. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

tagResources_tags :: Lens' TagResources (HashMap Text Text)

Specifies a list of tags that you want to add to the specified resources. A tag consists of a key and a value that you define.

Destructuring the Response

data TagResourcesResponse

See: newTagResourcesResponse smart constructor.

Constructors

TagResourcesResponse'

Fields

failedResourcesMap :: Maybe (HashMap Text FailureInfo)

A map containing a key-value pair for each failed item that couldn't be tagged. The key is the ARN of the failed resource. The value is a FailureInfo object that contains an error code, a status code, and an error message. If there are no errors, the FailedResourcesMap is empty.

httpStatus :: Int

The response's http status code.

newTagResourcesResponse

Arguments

:: Int
$sel:httpStatus:TagResourcesResponse'

-> TagResourcesResponse

Create a value of TagResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedResourcesMap:TagResourcesResponse', tagResourcesResponse_failedResourcesMap - A map containing a key-value pair for each failed item that couldn't be tagged. The key is the ARN of the failed resource. The value is a FailureInfo object that contains an error code, a status code, and an error message. If there are no errors, the FailedResourcesMap is empty.

$sel:httpStatus:TagResourcesResponse', tagResourcesResponse_httpStatus - The response's http status code.

Response Lenses

tagResourcesResponse_failedResourcesMap :: Lens' TagResourcesResponse (Maybe (HashMap Text FailureInfo))

A map containing a key-value pair for each failed item that couldn't be tagged. The key is the ARN of the failed resource. The value is a FailureInfo object that contains an error code, a status code, and an error message. If there are no errors, the FailedResourcesMap is empty.

tagResourcesResponse_httpStatus :: Lens' TagResourcesResponse Int

The response's http status code.