Amazonka.ResourceGroupsTagging.Types.ResourceTagMapping

Description

Documentation

data ResourceTagMapping

A list of resource ARNs and the tags (keys and values) that are associated with each.

See: newResourceTagMapping smart constructor.

Constructors

ResourceTagMapping'

Fields

complianceDetails :: Maybe ComplianceDetails

Information that shows whether a resource is compliant with the effective tag policy, including details on any noncompliant tag keys.

resourceARN :: Maybe Text

The ARN of the resource.

tags :: Maybe [Tag]

The tags that have been applied to one or more AWS resources.

newResourceTagMapping :: ResourceTagMapping

Create a value of ResourceTagMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:complianceDetails:ResourceTagMapping', resourceTagMapping_complianceDetails - Information that shows whether a resource is compliant with the effective tag policy, including details on any noncompliant tag keys.

$sel:resourceARN:ResourceTagMapping', resourceTagMapping_resourceARN - The ARN of the resource.

$sel:tags:ResourceTagMapping', resourceTagMapping_tags - The tags that have been applied to one or more AWS resources.

resourceTagMapping_complianceDetails :: Lens' ResourceTagMapping (Maybe ComplianceDetails)

Information that shows whether a resource is compliant with the effective tag policy, including details on any noncompliant tag keys.

resourceTagMapping_resourceARN :: Lens' ResourceTagMapping (Maybe Text)

The ARN of the resource.

resourceTagMapping_tags :: Lens' ResourceTagMapping (Maybe [Tag])

The tags that have been applied to one or more AWS resources.