Amazonka.ResourceGroupsTagging.Types.Summary

Description

Documentation

data Summary

A count of noncompliant resources.

See: newSummary smart constructor.

Constructors

Summary'

Fields

targetId :: Maybe Text

The account identifier or the root identifier of the organization. If you don't know the root ID, you can call the AWS Organizations ListRoots API.

lastUpdated :: Maybe Text

The timestamp that shows when this summary was generated in this Region.

resourceType :: Maybe Text

The AWS resource type.

nonCompliantResources :: Maybe Integer

The count of noncompliant resources.

targetIdType :: Maybe TargetIdType

Whether the target is an account, an OU, or the organization root.

region :: Maybe Text

The AWS Region that the summary applies to.

newSummary :: Summary

Create a value of Summary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetId:Summary', summary_targetId - The account identifier or the root identifier of the organization. If you don't know the root ID, you can call the AWS Organizations ListRoots API.

$sel:lastUpdated:Summary', summary_lastUpdated - The timestamp that shows when this summary was generated in this Region.

$sel:resourceType:Summary', summary_resourceType - The AWS resource type.

$sel:nonCompliantResources:Summary', summary_nonCompliantResources - The count of noncompliant resources.

$sel:targetIdType:Summary', summary_targetIdType - Whether the target is an account, an OU, or the organization root.

$sel:region:Summary', summary_region - The AWS Region that the summary applies to.

summary_targetId :: Lens' Summary (Maybe Text)

The account identifier or the root identifier of the organization. If you don't know the root ID, you can call the AWS Organizations ListRoots API.

summary_lastUpdated :: Lens' Summary (Maybe Text)

The timestamp that shows when this summary was generated in this Region.

summary_resourceType :: Lens' Summary (Maybe Text)

The AWS resource type.

summary_nonCompliantResources :: Lens' Summary (Maybe Integer)

The count of noncompliant resources.

summary_targetIdType :: Lens' Summary (Maybe TargetIdType)

Whether the target is an account, an OU, or the organization root.

summary_region :: Lens' Summary (Maybe Text)

The AWS Region that the summary applies to.