Amazonka.RobOMaker.BatchDescribeSimulationJob

Description

Describes one or more simulation jobs.

Creating a Request

data BatchDescribeSimulationJob

See: newBatchDescribeSimulationJob smart constructor.

Constructors

BatchDescribeSimulationJob'

Fields

jobs :: NonEmpty Text

A list of Amazon Resource Names (ARNs) of simulation jobs to describe.

newBatchDescribeSimulationJob

Arguments

:: NonEmpty Text
$sel:jobs:BatchDescribeSimulationJob'

-> BatchDescribeSimulationJob

Create a value of BatchDescribeSimulationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobs:BatchDescribeSimulationJob', batchDescribeSimulationJob_jobs - A list of Amazon Resource Names (ARNs) of simulation jobs to describe.

Request Lenses

batchDescribeSimulationJob_jobs :: Lens' BatchDescribeSimulationJob (NonEmpty Text)

A list of Amazon Resource Names (ARNs) of simulation jobs to describe.

Destructuring the Response

data BatchDescribeSimulationJobResponse

See: newBatchDescribeSimulationJobResponse smart constructor.

Constructors

BatchDescribeSimulationJobResponse'

Fields

unprocessedJobs :: Maybe (NonEmpty Text)

A list of unprocessed simulation job Amazon Resource Names (ARNs).

jobs :: Maybe [SimulationJob]

A list of simulation jobs.

httpStatus :: Int

The response's http status code.

newBatchDescribeSimulationJobResponse

Arguments

:: Int
$sel:httpStatus:BatchDescribeSimulationJobResponse'

-> BatchDescribeSimulationJobResponse

Create a value of BatchDescribeSimulationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unprocessedJobs:BatchDescribeSimulationJobResponse', batchDescribeSimulationJobResponse_unprocessedJobs - A list of unprocessed simulation job Amazon Resource Names (ARNs).

$sel:jobs:BatchDescribeSimulationJob', batchDescribeSimulationJobResponse_jobs - A list of simulation jobs.

$sel:httpStatus:BatchDescribeSimulationJobResponse', batchDescribeSimulationJobResponse_httpStatus - The response's http status code.

Response Lenses

batchDescribeSimulationJobResponse_unprocessedJobs :: Lens' BatchDescribeSimulationJobResponse (Maybe (NonEmpty Text))

A list of unprocessed simulation job Amazon Resource Names (ARNs).

batchDescribeSimulationJobResponse_jobs :: Lens' BatchDescribeSimulationJobResponse (Maybe [SimulationJob])

A list of simulation jobs.

batchDescribeSimulationJobResponse_httpStatus :: Lens' BatchDescribeSimulationJobResponse Int

The response's http status code.