Amazonka.RobOMaker.CreateDeploymentJob

Description

Deploys a specific version of a robot application to robots in a fleet.

The robot application must have a numbered applicationVersion for consistency reasons. To create a new version, use CreateRobotApplicationVersion or see Creating a Robot Application Version.

After 90 days, deployment jobs expire and will be deleted. They will no longer be accessible.

Creating a Request

data CreateDeploymentJob

See: newCreateDeploymentJob smart constructor.

Constructors

CreateDeploymentJob'

Fields

deploymentConfig :: Maybe DeploymentConfig

The requested deployment configuration.

tags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the deployment job.

clientRequestToken :: Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

fleet :: Text

The Amazon Resource Name (ARN) of the fleet to deploy.

deploymentApplicationConfigs :: NonEmpty DeploymentApplicationConfig

The deployment application configuration.

newCreateDeploymentJob

Arguments

:: Text
$sel:clientRequestToken:CreateDeploymentJob'

-> Text
$sel:fleet:CreateDeploymentJob'

-> NonEmpty DeploymentApplicationConfig
$sel:deploymentApplicationConfigs:CreateDeploymentJob'

-> CreateDeploymentJob

Create a value of CreateDeploymentJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentConfig:CreateDeploymentJob', createDeploymentJob_deploymentConfig - The requested deployment configuration.

$sel:tags:CreateDeploymentJob', createDeploymentJob_tags - A map that contains tag keys and tag values that are attached to the deployment job.

$sel:clientRequestToken:CreateDeploymentJob', createDeploymentJob_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:fleet:CreateDeploymentJob', createDeploymentJob_fleet - The Amazon Resource Name (ARN) of the fleet to deploy.

$sel:deploymentApplicationConfigs:CreateDeploymentJob', createDeploymentJob_deploymentApplicationConfigs - The deployment application configuration.

Request Lenses

createDeploymentJob_deploymentConfig :: Lens' CreateDeploymentJob (Maybe DeploymentConfig)

The requested deployment configuration.

createDeploymentJob_tags :: Lens' CreateDeploymentJob (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the deployment job.

createDeploymentJob_clientRequestToken :: Lens' CreateDeploymentJob Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createDeploymentJob_fleet :: Lens' CreateDeploymentJob Text

The Amazon Resource Name (ARN) of the fleet to deploy.

createDeploymentJob_deploymentApplicationConfigs :: Lens' CreateDeploymentJob (NonEmpty DeploymentApplicationConfig)

The deployment application configuration.

Destructuring the Response

data CreateDeploymentJobResponse

See: newCreateDeploymentJobResponse smart constructor.

Constructors

CreateDeploymentJobResponse'

Fields

failureReason :: Maybe Text

The failure reason of the deployment job if it failed.

status :: Maybe DeploymentStatus

The status of the deployment job.

deploymentApplicationConfigs :: Maybe (NonEmpty DeploymentApplicationConfig)

The deployment application configuration.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the deployment job.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the fleet was created.

failureCode :: Maybe DeploymentJobErrorCode

The failure code of the simulation job if it failed:

BadPermissionError

AWS Greengrass requires a service-level role permission to access other services. The role must include the AWSGreengrassResourceAccessRolePolicy managed policy.

ExtractingBundleFailure

The robot application could not be extracted from the bundle.

FailureThresholdBreached

The percentage of robots that could not be updated exceeded the percentage set for the deployment.

GreengrassDeploymentFailed

The robot application could not be deployed to the robot.

GreengrassGroupVersionDoesNotExist

The AWS Greengrass group or version associated with a robot is missing.

InternalServerError

An internal error has occurred. Retry your request, but if the problem persists, contact us with details.

MissingRobotApplicationArchitecture

The robot application does not have a source that matches the architecture of the robot.

MissingRobotDeploymentResource

One or more of the resources specified for the robot application are missing. For example, does the robot application have the correct launch package and launch file?

PostLaunchFileFailure

The post-launch script failed.

PreLaunchFileFailure

The pre-launch script failed.

ResourceNotFound

One or more deployment resources are missing. For example, do robot application source bundles still exist?

RobotDeploymentNoResponse

There is no response from the robot. It might not be powered on or connected to the internet.

deploymentConfig :: Maybe DeploymentConfig

The deployment configuration.

fleet :: Maybe Text

The target fleet for the deployment job.

tags :: Maybe (HashMap Text Text)

The list of all tags added to the deployment job.

httpStatus :: Int

The response's http status code.

newCreateDeploymentJobResponse

Arguments

:: Int
$sel:httpStatus:CreateDeploymentJobResponse'

-> CreateDeploymentJobResponse

Create a value of CreateDeploymentJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:CreateDeploymentJobResponse', createDeploymentJobResponse_failureReason - The failure reason of the deployment job if it failed.

$sel:status:CreateDeploymentJobResponse', createDeploymentJobResponse_status - The status of the deployment job.

$sel:deploymentApplicationConfigs:CreateDeploymentJob', createDeploymentJobResponse_deploymentApplicationConfigs - The deployment application configuration.

$sel:arn:CreateDeploymentJobResponse', createDeploymentJobResponse_arn - The Amazon Resource Name (ARN) of the deployment job.

$sel:createdAt:CreateDeploymentJobResponse', createDeploymentJobResponse_createdAt - The time, in milliseconds since the epoch, when the fleet was created.

$sel:failureCode:CreateDeploymentJobResponse', createDeploymentJobResponse_failureCode - The failure code of the simulation job if it failed:

BadPermissionError

AWS Greengrass requires a service-level role permission to access other services. The role must include the AWSGreengrassResourceAccessRolePolicy managed policy.

ExtractingBundleFailure

The robot application could not be extracted from the bundle.

FailureThresholdBreached

The percentage of robots that could not be updated exceeded the percentage set for the deployment.

GreengrassDeploymentFailed

The robot application could not be deployed to the robot.

GreengrassGroupVersionDoesNotExist

The AWS Greengrass group or version associated with a robot is missing.

InternalServerError

An internal error has occurred. Retry your request, but if the problem persists, contact us with details.

MissingRobotApplicationArchitecture

The robot application does not have a source that matches the architecture of the robot.

MissingRobotDeploymentResource

One or more of the resources specified for the robot application are missing. For example, does the robot application have the correct launch package and launch file?

PostLaunchFileFailure

The post-launch script failed.

PreLaunchFileFailure

The pre-launch script failed.

ResourceNotFound

One or more deployment resources are missing. For example, do robot application source bundles still exist?

RobotDeploymentNoResponse

There is no response from the robot. It might not be powered on or connected to the internet.

$sel:deploymentConfig:CreateDeploymentJob', createDeploymentJobResponse_deploymentConfig - The deployment configuration.

$sel:fleet:CreateDeploymentJob', createDeploymentJobResponse_fleet - The target fleet for the deployment job.

$sel:tags:CreateDeploymentJob', createDeploymentJobResponse_tags - The list of all tags added to the deployment job.

$sel:httpStatus:CreateDeploymentJobResponse', createDeploymentJobResponse_httpStatus - The response's http status code.

Response Lenses

createDeploymentJobResponse_failureReason :: Lens' CreateDeploymentJobResponse (Maybe Text)

The failure reason of the deployment job if it failed.

createDeploymentJobResponse_status :: Lens' CreateDeploymentJobResponse (Maybe DeploymentStatus)

The status of the deployment job.

createDeploymentJobResponse_deploymentApplicationConfigs :: Lens' CreateDeploymentJobResponse (Maybe (NonEmpty DeploymentApplicationConfig))

The deployment application configuration.

createDeploymentJobResponse_arn :: Lens' CreateDeploymentJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the deployment job.

createDeploymentJobResponse_createdAt :: Lens' CreateDeploymentJobResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the fleet was created.

createDeploymentJobResponse_failureCode :: Lens' CreateDeploymentJobResponse (Maybe DeploymentJobErrorCode)

The failure code of the simulation job if it failed:

BadPermissionError

AWS Greengrass requires a service-level role permission to access other services. The role must include the AWSGreengrassResourceAccessRolePolicy managed policy.

ExtractingBundleFailure

The robot application could not be extracted from the bundle.

FailureThresholdBreached

The percentage of robots that could not be updated exceeded the percentage set for the deployment.

GreengrassDeploymentFailed

The robot application could not be deployed to the robot.

GreengrassGroupVersionDoesNotExist

The AWS Greengrass group or version associated with a robot is missing.

InternalServerError

An internal error has occurred. Retry your request, but if the problem persists, contact us with details.

MissingRobotApplicationArchitecture

The robot application does not have a source that matches the architecture of the robot.

MissingRobotDeploymentResource

One or more of the resources specified for the robot application are missing. For example, does the robot application have the correct launch package and launch file?

PostLaunchFileFailure

The post-launch script failed.

PreLaunchFileFailure

The pre-launch script failed.

ResourceNotFound

One or more deployment resources are missing. For example, do robot application source bundles still exist?

RobotDeploymentNoResponse

There is no response from the robot. It might not be powered on or connected to the internet.

createDeploymentJobResponse_deploymentConfig :: Lens' CreateDeploymentJobResponse (Maybe DeploymentConfig)

The deployment configuration.

createDeploymentJobResponse_fleet :: Lens' CreateDeploymentJobResponse (Maybe Text)

The target fleet for the deployment job.

createDeploymentJobResponse_tags :: Lens' CreateDeploymentJobResponse (Maybe (HashMap Text Text))

The list of all tags added to the deployment job.

createDeploymentJobResponse_httpStatus :: Lens' CreateDeploymentJobResponse Int

The response's http status code.