Amazonka.RobOMaker.CreateRobotApplication

Description

Creates a robot application.

Creating a Request

data CreateRobotApplication

See: newCreateRobotApplication smart constructor.

Constructors

CreateRobotApplication'

Fields

environment :: Maybe Environment

The object that contains that URI of the Docker image that you use for your robot application.

sources :: Maybe [SourceConfig]

The sources of the robot application.

tags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the robot application.

name :: Text

The name of the robot application.

robotSoftwareSuite :: RobotSoftwareSuite

The robot software suite (ROS distribuition) used by the robot application.

newCreateRobotApplication

Arguments

:: Text
$sel:name:CreateRobotApplication'

-> RobotSoftwareSuite
$sel:robotSoftwareSuite:CreateRobotApplication'

-> CreateRobotApplication

Create a value of CreateRobotApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environment:CreateRobotApplication', createRobotApplication_environment - The object that contains that URI of the Docker image that you use for your robot application.

$sel:sources:CreateRobotApplication', createRobotApplication_sources - The sources of the robot application.

$sel:tags:CreateRobotApplication', createRobotApplication_tags - A map that contains tag keys and tag values that are attached to the robot application.

$sel:name:CreateRobotApplication', createRobotApplication_name - The name of the robot application.

$sel:robotSoftwareSuite:CreateRobotApplication', createRobotApplication_robotSoftwareSuite - The robot software suite (ROS distribuition) used by the robot application.

Request Lenses

createRobotApplication_environment :: Lens' CreateRobotApplication (Maybe Environment)

The object that contains that URI of the Docker image that you use for your robot application.

createRobotApplication_sources :: Lens' CreateRobotApplication (Maybe [SourceConfig])

The sources of the robot application.

createRobotApplication_tags :: Lens' CreateRobotApplication (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the robot application.

createRobotApplication_name :: Lens' CreateRobotApplication Text

The name of the robot application.

createRobotApplication_robotSoftwareSuite :: Lens' CreateRobotApplication RobotSoftwareSuite

The robot software suite (ROS distribuition) used by the robot application.

Destructuring the Response

data CreateRobotApplicationResponse

See: newCreateRobotApplicationResponse smart constructor.

Constructors

CreateRobotApplicationResponse'

Fields

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the robot application was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the robot application.

environment :: Maybe Environment

An object that contains the Docker image URI used to a create your robot application.

sources :: Maybe [Source]

The sources of the robot application.

name :: Maybe Text

The name of the robot application.

version :: Maybe Text

The version of the robot application.

robotSoftwareSuite :: Maybe RobotSoftwareSuite

The robot software suite (ROS distribution) used by the robot application.

revisionId :: Maybe Text

The revision id of the robot application.

tags :: Maybe (HashMap Text Text)

The list of all tags added to the robot application.

httpStatus :: Int

The response's http status code.

newCreateRobotApplicationResponse

Arguments

:: Int
$sel:httpStatus:CreateRobotApplicationResponse'

-> CreateRobotApplicationResponse

Create a value of CreateRobotApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedAt:CreateRobotApplicationResponse', createRobotApplicationResponse_lastUpdatedAt - The time, in milliseconds since the epoch, when the robot application was last updated.

$sel:arn:CreateRobotApplicationResponse', createRobotApplicationResponse_arn - The Amazon Resource Name (ARN) of the robot application.

$sel:environment:CreateRobotApplication', createRobotApplicationResponse_environment - An object that contains the Docker image URI used to a create your robot application.

$sel:sources:CreateRobotApplication', createRobotApplicationResponse_sources - The sources of the robot application.

$sel:name:CreateRobotApplication', createRobotApplicationResponse_name - The name of the robot application.

$sel:version:CreateRobotApplicationResponse', createRobotApplicationResponse_version - The version of the robot application.

$sel:robotSoftwareSuite:CreateRobotApplication', createRobotApplicationResponse_robotSoftwareSuite - The robot software suite (ROS distribution) used by the robot application.

$sel:revisionId:CreateRobotApplicationResponse', createRobotApplicationResponse_revisionId - The revision id of the robot application.

$sel:tags:CreateRobotApplication', createRobotApplicationResponse_tags - The list of all tags added to the robot application.

$sel:httpStatus:CreateRobotApplicationResponse', createRobotApplicationResponse_httpStatus - The response's http status code.

Response Lenses

createRobotApplicationResponse_lastUpdatedAt :: Lens' CreateRobotApplicationResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the robot application was last updated.

createRobotApplicationResponse_arn :: Lens' CreateRobotApplicationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the robot application.

createRobotApplicationResponse_environment :: Lens' CreateRobotApplicationResponse (Maybe Environment)

An object that contains the Docker image URI used to a create your robot application.

createRobotApplicationResponse_sources :: Lens' CreateRobotApplicationResponse (Maybe [Source])

The sources of the robot application.

createRobotApplicationResponse_name :: Lens' CreateRobotApplicationResponse (Maybe Text)

The name of the robot application.

createRobotApplicationResponse_version :: Lens' CreateRobotApplicationResponse (Maybe Text)

The version of the robot application.

createRobotApplicationResponse_robotSoftwareSuite :: Lens' CreateRobotApplicationResponse (Maybe RobotSoftwareSuite)

The robot software suite (ROS distribution) used by the robot application.

createRobotApplicationResponse_revisionId :: Lens' CreateRobotApplicationResponse (Maybe Text)

The revision id of the robot application.

createRobotApplicationResponse_tags :: Lens' CreateRobotApplicationResponse (Maybe (HashMap Text Text))

The list of all tags added to the robot application.

createRobotApplicationResponse_httpStatus :: Lens' CreateRobotApplicationResponse Int

The response's http status code.