Amazonka.RobOMaker.CreateSimulationApplication

Description

Creates a simulation application.

Creating a Request

data CreateSimulationApplication

See: newCreateSimulationApplication smart constructor.

Constructors

CreateSimulationApplication'

Fields

renderingEngine :: Maybe RenderingEngine

The rendering engine for the simulation application.

environment :: Maybe Environment

The object that contains the Docker image URI used to create your simulation application.

sources :: Maybe [SourceConfig]

The sources of the simulation application.

tags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the simulation application.

name :: Text

The name of the simulation application.

simulationSoftwareSuite :: SimulationSoftwareSuite

The simulation software suite used by the simulation application.

robotSoftwareSuite :: RobotSoftwareSuite

The robot software suite (ROS distribution) used by the simulation application.

newCreateSimulationApplication

Arguments

:: Text
$sel:name:CreateSimulationApplication'

-> SimulationSoftwareSuite
$sel:simulationSoftwareSuite:CreateSimulationApplication'

-> RobotSoftwareSuite
$sel:robotSoftwareSuite:CreateSimulationApplication'

-> CreateSimulationApplication

Create a value of CreateSimulationApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renderingEngine:CreateSimulationApplication', createSimulationApplication_renderingEngine - The rendering engine for the simulation application.

$sel:environment:CreateSimulationApplication', createSimulationApplication_environment - The object that contains the Docker image URI used to create your simulation application.

$sel:sources:CreateSimulationApplication', createSimulationApplication_sources - The sources of the simulation application.

$sel:tags:CreateSimulationApplication', createSimulationApplication_tags - A map that contains tag keys and tag values that are attached to the simulation application.

$sel:name:CreateSimulationApplication', createSimulationApplication_name - The name of the simulation application.

$sel:simulationSoftwareSuite:CreateSimulationApplication', createSimulationApplication_simulationSoftwareSuite - The simulation software suite used by the simulation application.

$sel:robotSoftwareSuite:CreateSimulationApplication', createSimulationApplication_robotSoftwareSuite - The robot software suite (ROS distribution) used by the simulation application.

Request Lenses

createSimulationApplication_renderingEngine :: Lens' CreateSimulationApplication (Maybe RenderingEngine)

The rendering engine for the simulation application.

createSimulationApplication_environment :: Lens' CreateSimulationApplication (Maybe Environment)

The object that contains the Docker image URI used to create your simulation application.

createSimulationApplication_sources :: Lens' CreateSimulationApplication (Maybe [SourceConfig])

The sources of the simulation application.

createSimulationApplication_tags :: Lens' CreateSimulationApplication (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the simulation application.

createSimulationApplication_name :: Lens' CreateSimulationApplication Text

The name of the simulation application.

createSimulationApplication_simulationSoftwareSuite :: Lens' CreateSimulationApplication SimulationSoftwareSuite

The simulation software suite used by the simulation application.

createSimulationApplication_robotSoftwareSuite :: Lens' CreateSimulationApplication RobotSoftwareSuite

The robot software suite (ROS distribution) used by the simulation application.

Destructuring the Response

data CreateSimulationApplicationResponse

See: newCreateSimulationApplicationResponse smart constructor.

Constructors

CreateSimulationApplicationResponse'

Fields

renderingEngine :: Maybe RenderingEngine

The rendering engine for the simulation application.

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation application was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the simulation application.

environment :: Maybe Environment

The object that contains the Docker image URI that you used to create your simulation application.

sources :: Maybe [Source]

The sources of the simulation application.

name :: Maybe Text

The name of the simulation application.

version :: Maybe Text

The version of the simulation application.

simulationSoftwareSuite :: Maybe SimulationSoftwareSuite

The simulation software suite used by the simulation application.

robotSoftwareSuite :: Maybe RobotSoftwareSuite

Information about the robot software suite (ROS distribution).

revisionId :: Maybe Text

The revision id of the simulation application.

tags :: Maybe (HashMap Text Text)

The list of all tags added to the simulation application.

httpStatus :: Int

The response's http status code.

newCreateSimulationApplicationResponse

Arguments

:: Int
$sel:httpStatus:CreateSimulationApplicationResponse'

-> CreateSimulationApplicationResponse

Create a value of CreateSimulationApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renderingEngine:CreateSimulationApplication', createSimulationApplicationResponse_renderingEngine - The rendering engine for the simulation application.

$sel:lastUpdatedAt:CreateSimulationApplicationResponse', createSimulationApplicationResponse_lastUpdatedAt - The time, in milliseconds since the epoch, when the simulation application was last updated.

$sel:arn:CreateSimulationApplicationResponse', createSimulationApplicationResponse_arn - The Amazon Resource Name (ARN) of the simulation application.

$sel:environment:CreateSimulationApplication', createSimulationApplicationResponse_environment - The object that contains the Docker image URI that you used to create your simulation application.

$sel:sources:CreateSimulationApplication', createSimulationApplicationResponse_sources - The sources of the simulation application.

$sel:name:CreateSimulationApplication', createSimulationApplicationResponse_name - The name of the simulation application.

$sel:version:CreateSimulationApplicationResponse', createSimulationApplicationResponse_version - The version of the simulation application.

$sel:simulationSoftwareSuite:CreateSimulationApplication', createSimulationApplicationResponse_simulationSoftwareSuite - The simulation software suite used by the simulation application.

$sel:robotSoftwareSuite:CreateSimulationApplication', createSimulationApplicationResponse_robotSoftwareSuite - Information about the robot software suite (ROS distribution).

$sel:revisionId:CreateSimulationApplicationResponse', createSimulationApplicationResponse_revisionId - The revision id of the simulation application.

$sel:tags:CreateSimulationApplication', createSimulationApplicationResponse_tags - The list of all tags added to the simulation application.

$sel:httpStatus:CreateSimulationApplicationResponse', createSimulationApplicationResponse_httpStatus - The response's http status code.

Response Lenses

createSimulationApplicationResponse_renderingEngine :: Lens' CreateSimulationApplicationResponse (Maybe RenderingEngine)

The rendering engine for the simulation application.

createSimulationApplicationResponse_lastUpdatedAt :: Lens' CreateSimulationApplicationResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation application was last updated.

createSimulationApplicationResponse_arn :: Lens' CreateSimulationApplicationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the simulation application.

createSimulationApplicationResponse_environment :: Lens' CreateSimulationApplicationResponse (Maybe Environment)

The object that contains the Docker image URI that you used to create your simulation application.

createSimulationApplicationResponse_sources :: Lens' CreateSimulationApplicationResponse (Maybe [Source])

The sources of the simulation application.

createSimulationApplicationResponse_name :: Lens' CreateSimulationApplicationResponse (Maybe Text)

The name of the simulation application.

createSimulationApplicationResponse_version :: Lens' CreateSimulationApplicationResponse (Maybe Text)

The version of the simulation application.

createSimulationApplicationResponse_simulationSoftwareSuite :: Lens' CreateSimulationApplicationResponse (Maybe SimulationSoftwareSuite)

The simulation software suite used by the simulation application.

createSimulationApplicationResponse_robotSoftwareSuite :: Lens' CreateSimulationApplicationResponse (Maybe RobotSoftwareSuite)

Information about the robot software suite (ROS distribution).

createSimulationApplicationResponse_revisionId :: Lens' CreateSimulationApplicationResponse (Maybe Text)

The revision id of the simulation application.

createSimulationApplicationResponse_tags :: Lens' CreateSimulationApplicationResponse (Maybe (HashMap Text Text))

The list of all tags added to the simulation application.

createSimulationApplicationResponse_httpStatus :: Lens' CreateSimulationApplicationResponse Int

The response's http status code.