Amazonka.RobOMaker.CreateSimulationJob

Description

Creates a simulation job.

After 90 days, simulation jobs expire and will be deleted. They will no longer be accessible.

Creating a Request

data CreateSimulationJob

See: newCreateSimulationJob smart constructor.

Constructors

CreateSimulationJob'

Fields

failureBehavior :: Maybe FailureBehavior

The failure behavior the simulation job.

Continue

Leaves the instance running for its maximum timeout duration after a 4XX error code.

Fail

Stop the simulation job and terminate the instance.

robotApplications :: Maybe (NonEmpty RobotApplicationConfig)

The robot application to use in the simulation job.

compute :: Maybe Compute

Compute information for the simulation job.

dataSources :: Maybe (NonEmpty DataSourceConfig)

Specify data sources to mount read-only files from S3 into your simulation. These files are available under /opt/robomaker/datasources/data_source_name.

There is a limit of 100 files and a combined size of 25GB for all DataSourceConfig objects.

vpcConfig :: Maybe VPCConfig

If your simulation job accesses resources in a VPC, you provide this parameter identifying the list of security group IDs and subnet IDs. These must belong to the same VPC. You must provide at least one security group and one subnet ID.

outputLocation :: Maybe OutputLocation

Location for output files generated by the simulation job.

simulationApplications :: Maybe (NonEmpty SimulationApplicationConfig)

The simulation application to use in the simulation job.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

loggingConfig :: Maybe LoggingConfig

The logging configuration.

tags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the simulation job.

maxJobDurationInSeconds :: Integer

The maximum simulation job duration in seconds (up to 14 days or 1,209,600 seconds. When maxJobDurationInSeconds is reached, the simulation job will status will transition to Completed.

iamRole :: Text

The IAM role name that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf. This is how credentials are passed in to your simulation job.

newCreateSimulationJob

Arguments

:: Integer
$sel:maxJobDurationInSeconds:CreateSimulationJob'

-> Text
$sel:iamRole:CreateSimulationJob'

-> CreateSimulationJob

Create a value of CreateSimulationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureBehavior:CreateSimulationJob', createSimulationJob_failureBehavior - The failure behavior the simulation job.

Continue

Leaves the instance running for its maximum timeout duration after a 4XX error code.

Fail

Stop the simulation job and terminate the instance.

$sel:robotApplications:CreateSimulationJob', createSimulationJob_robotApplications - The robot application to use in the simulation job.

$sel:compute:CreateSimulationJob', createSimulationJob_compute - Compute information for the simulation job.

$sel:dataSources:CreateSimulationJob', createSimulationJob_dataSources - Specify data sources to mount read-only files from S3 into your simulation. These files are available under /opt/robomaker/datasources/data_source_name.

There is a limit of 100 files and a combined size of 25GB for all DataSourceConfig objects.

$sel:vpcConfig:CreateSimulationJob', createSimulationJob_vpcConfig - If your simulation job accesses resources in a VPC, you provide this parameter identifying the list of security group IDs and subnet IDs. These must belong to the same VPC. You must provide at least one security group and one subnet ID.

$sel:outputLocation:CreateSimulationJob', createSimulationJob_outputLocation - Location for output files generated by the simulation job.

$sel:simulationApplications:CreateSimulationJob', createSimulationJob_simulationApplications - The simulation application to use in the simulation job.

$sel:clientRequestToken:CreateSimulationJob', createSimulationJob_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:loggingConfig:CreateSimulationJob', createSimulationJob_loggingConfig - The logging configuration.

$sel:tags:CreateSimulationJob', createSimulationJob_tags - A map that contains tag keys and tag values that are attached to the simulation job.

$sel:maxJobDurationInSeconds:CreateSimulationJob', createSimulationJob_maxJobDurationInSeconds - The maximum simulation job duration in seconds (up to 14 days or 1,209,600 seconds. When maxJobDurationInSeconds is reached, the simulation job will status will transition to Completed.

$sel:iamRole:CreateSimulationJob', createSimulationJob_iamRole - The IAM role name that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf. This is how credentials are passed in to your simulation job.

Request Lenses

createSimulationJob_failureBehavior :: Lens' CreateSimulationJob (Maybe FailureBehavior)

The failure behavior the simulation job.

Continue

Leaves the instance running for its maximum timeout duration after a 4XX error code.

Fail

Stop the simulation job and terminate the instance.

createSimulationJob_robotApplications :: Lens' CreateSimulationJob (Maybe (NonEmpty RobotApplicationConfig))

The robot application to use in the simulation job.

createSimulationJob_compute :: Lens' CreateSimulationJob (Maybe Compute)

Compute information for the simulation job.

createSimulationJob_dataSources :: Lens' CreateSimulationJob (Maybe (NonEmpty DataSourceConfig))

Specify data sources to mount read-only files from S3 into your simulation. These files are available under /opt/robomaker/datasources/data_source_name.

There is a limit of 100 files and a combined size of 25GB for all DataSourceConfig objects.

createSimulationJob_vpcConfig :: Lens' CreateSimulationJob (Maybe VPCConfig)

If your simulation job accesses resources in a VPC, you provide this parameter identifying the list of security group IDs and subnet IDs. These must belong to the same VPC. You must provide at least one security group and one subnet ID.

createSimulationJob_outputLocation :: Lens' CreateSimulationJob (Maybe OutputLocation)

Location for output files generated by the simulation job.

createSimulationJob_simulationApplications :: Lens' CreateSimulationJob (Maybe (NonEmpty SimulationApplicationConfig))

The simulation application to use in the simulation job.

createSimulationJob_clientRequestToken :: Lens' CreateSimulationJob (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createSimulationJob_loggingConfig :: Lens' CreateSimulationJob (Maybe LoggingConfig)

The logging configuration.

createSimulationJob_tags :: Lens' CreateSimulationJob (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the simulation job.

createSimulationJob_maxJobDurationInSeconds :: Lens' CreateSimulationJob Integer

The maximum simulation job duration in seconds (up to 14 days or 1,209,600 seconds. When maxJobDurationInSeconds is reached, the simulation job will status will transition to Completed.

createSimulationJob_iamRole :: Lens' CreateSimulationJob Text

The IAM role name that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf. This is how credentials are passed in to your simulation job.

Destructuring the Response

data CreateSimulationJobResponse

See: newCreateSimulationJobResponse smart constructor.

Constructors

CreateSimulationJobResponse'

Fields

failureBehavior :: Maybe FailureBehavior

the failure behavior for the simulation job.

status :: Maybe SimulationJobStatus

The status of the simulation job.

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation job was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the simulation job.

robotApplications :: Maybe (NonEmpty RobotApplicationConfig)

The robot application used by the simulation job.

failureCode :: Maybe SimulationJobErrorCode

The failure code of the simulation job if it failed:

InternalServiceError

Internal service error.

RobotApplicationCrash

Robot application exited abnormally.

SimulationApplicationCrash

Simulation application exited abnormally.

BadPermissionsRobotApplication

Robot application bundle could not be downloaded.

BadPermissionsSimulationApplication

Simulation application bundle could not be downloaded.

BadPermissionsS3Output

Unable to publish outputs to customer-provided S3 bucket.

BadPermissionsCloudwatchLogs

Unable to publish logs to customer-provided CloudWatch Logs resource.

SubnetIpLimitExceeded

Subnet IP limit exceeded.

ENILimitExceeded

ENI limit exceeded.

BadPermissionsUserCredentials

Unable to use the Role provided.

InvalidBundleRobotApplication

Robot bundle cannot be extracted (invalid format, bundling error, or other issue).

InvalidBundleSimulationApplication

Simulation bundle cannot be extracted (invalid format, bundling error, or other issue).

RobotApplicationVersionMismatchedEtag

Etag for RobotApplication does not match value during version creation.

SimulationApplicationVersionMismatchedEtag

Etag for SimulationApplication does not match value during version creation.

compute :: Maybe ComputeResponse

Compute information for the simulation job.

dataSources :: Maybe [DataSource]

The data sources for the simulation job.

vpcConfig :: Maybe VPCConfigResponse

Information about the vpc configuration.

outputLocation :: Maybe OutputLocation

Simulation job output files location.

simulationApplications :: Maybe (NonEmpty SimulationApplicationConfig)

The simulation application used by the simulation job.

simulationTimeMillis :: Maybe Integer

The simulation job execution duration in milliseconds.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

lastStartedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation job was last started.

loggingConfig :: Maybe LoggingConfig

The logging configuration.

iamRole :: Maybe Text

The IAM role that allows the simulation job to call the AWS APIs that are specified in its associated policies on your behalf.

maxJobDurationInSeconds :: Maybe Integer

The maximum simulation job duration in seconds.

tags :: Maybe (HashMap Text Text)

The list of all tags added to the simulation job.

httpStatus :: Int

The response's http status code.

newCreateSimulationJobResponse

Arguments

:: Int
$sel:httpStatus:CreateSimulationJobResponse'

-> CreateSimulationJobResponse

Create a value of CreateSimulationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureBehavior:CreateSimulationJob', createSimulationJobResponse_failureBehavior - the failure behavior for the simulation job.

$sel:status:CreateSimulationJobResponse', createSimulationJobResponse_status - The status of the simulation job.

$sel:lastUpdatedAt:CreateSimulationJobResponse', createSimulationJobResponse_lastUpdatedAt - The time, in milliseconds since the epoch, when the simulation job was last updated.

$sel:arn:CreateSimulationJobResponse', createSimulationJobResponse_arn - The Amazon Resource Name (ARN) of the simulation job.

$sel:robotApplications:CreateSimulationJob', createSimulationJobResponse_robotApplications - The robot application used by the simulation job.

$sel:failureCode:CreateSimulationJobResponse', createSimulationJobResponse_failureCode - The failure code of the simulation job if it failed:

InternalServiceError

Internal service error.

RobotApplicationCrash

Robot application exited abnormally.

SimulationApplicationCrash

Simulation application exited abnormally.

BadPermissionsRobotApplication

Robot application bundle could not be downloaded.

BadPermissionsSimulationApplication

Simulation application bundle could not be downloaded.

BadPermissionsS3Output

Unable to publish outputs to customer-provided S3 bucket.

BadPermissionsCloudwatchLogs

Unable to publish logs to customer-provided CloudWatch Logs resource.

SubnetIpLimitExceeded

Subnet IP limit exceeded.

ENILimitExceeded

ENI limit exceeded.

BadPermissionsUserCredentials

Unable to use the Role provided.

InvalidBundleRobotApplication

Robot bundle cannot be extracted (invalid format, bundling error, or other issue).

InvalidBundleSimulationApplication

Simulation bundle cannot be extracted (invalid format, bundling error, or other issue).

RobotApplicationVersionMismatchedEtag

Etag for RobotApplication does not match value during version creation.

SimulationApplicationVersionMismatchedEtag

Etag for SimulationApplication does not match value during version creation.

$sel:compute:CreateSimulationJob', createSimulationJobResponse_compute - Compute information for the simulation job.

$sel:dataSources:CreateSimulationJob', createSimulationJobResponse_dataSources - The data sources for the simulation job.

$sel:vpcConfig:CreateSimulationJob', createSimulationJobResponse_vpcConfig - Information about the vpc configuration.

$sel:outputLocation:CreateSimulationJob', createSimulationJobResponse_outputLocation - Simulation job output files location.

$sel:simulationApplications:CreateSimulationJob', createSimulationJobResponse_simulationApplications - The simulation application used by the simulation job.

$sel:simulationTimeMillis:CreateSimulationJobResponse', createSimulationJobResponse_simulationTimeMillis - The simulation job execution duration in milliseconds.

$sel:clientRequestToken:CreateSimulationJob', createSimulationJobResponse_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:lastStartedAt:CreateSimulationJobResponse', createSimulationJobResponse_lastStartedAt - The time, in milliseconds since the epoch, when the simulation job was last started.

$sel:loggingConfig:CreateSimulationJob', createSimulationJobResponse_loggingConfig - The logging configuration.

$sel:iamRole:CreateSimulationJob', createSimulationJobResponse_iamRole - The IAM role that allows the simulation job to call the AWS APIs that are specified in its associated policies on your behalf.

$sel:maxJobDurationInSeconds:CreateSimulationJob', createSimulationJobResponse_maxJobDurationInSeconds - The maximum simulation job duration in seconds.

$sel:tags:CreateSimulationJob', createSimulationJobResponse_tags - The list of all tags added to the simulation job.

$sel:httpStatus:CreateSimulationJobResponse', createSimulationJobResponse_httpStatus - The response's http status code.

Response Lenses

createSimulationJobResponse_failureBehavior :: Lens' CreateSimulationJobResponse (Maybe FailureBehavior)

the failure behavior for the simulation job.

createSimulationJobResponse_status :: Lens' CreateSimulationJobResponse (Maybe SimulationJobStatus)

The status of the simulation job.

createSimulationJobResponse_lastUpdatedAt :: Lens' CreateSimulationJobResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation job was last updated.

createSimulationJobResponse_arn :: Lens' CreateSimulationJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the simulation job.

createSimulationJobResponse_robotApplications :: Lens' CreateSimulationJobResponse (Maybe (NonEmpty RobotApplicationConfig))

The robot application used by the simulation job.

createSimulationJobResponse_failureCode :: Lens' CreateSimulationJobResponse (Maybe SimulationJobErrorCode)

The failure code of the simulation job if it failed:

InternalServiceError

Internal service error.

RobotApplicationCrash

Robot application exited abnormally.

SimulationApplicationCrash

Simulation application exited abnormally.

BadPermissionsRobotApplication

Robot application bundle could not be downloaded.

BadPermissionsSimulationApplication

Simulation application bundle could not be downloaded.

BadPermissionsS3Output

Unable to publish outputs to customer-provided S3 bucket.

BadPermissionsCloudwatchLogs

Unable to publish logs to customer-provided CloudWatch Logs resource.

SubnetIpLimitExceeded

Subnet IP limit exceeded.

ENILimitExceeded

ENI limit exceeded.

BadPermissionsUserCredentials

Unable to use the Role provided.

InvalidBundleRobotApplication

Robot bundle cannot be extracted (invalid format, bundling error, or other issue).

InvalidBundleSimulationApplication

Simulation bundle cannot be extracted (invalid format, bundling error, or other issue).

RobotApplicationVersionMismatchedEtag

Etag for RobotApplication does not match value during version creation.

SimulationApplicationVersionMismatchedEtag

Etag for SimulationApplication does not match value during version creation.

createSimulationJobResponse_compute :: Lens' CreateSimulationJobResponse (Maybe ComputeResponse)

Compute information for the simulation job.

createSimulationJobResponse_dataSources :: Lens' CreateSimulationJobResponse (Maybe [DataSource])

The data sources for the simulation job.

createSimulationJobResponse_vpcConfig :: Lens' CreateSimulationJobResponse (Maybe VPCConfigResponse)

Information about the vpc configuration.

createSimulationJobResponse_outputLocation :: Lens' CreateSimulationJobResponse (Maybe OutputLocation)

Simulation job output files location.

createSimulationJobResponse_simulationApplications :: Lens' CreateSimulationJobResponse (Maybe (NonEmpty SimulationApplicationConfig))

The simulation application used by the simulation job.

createSimulationJobResponse_simulationTimeMillis :: Lens' CreateSimulationJobResponse (Maybe Integer)

The simulation job execution duration in milliseconds.

createSimulationJobResponse_clientRequestToken :: Lens' CreateSimulationJobResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createSimulationJobResponse_lastStartedAt :: Lens' CreateSimulationJobResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation job was last started.

createSimulationJobResponse_loggingConfig :: Lens' CreateSimulationJobResponse (Maybe LoggingConfig)

The logging configuration.

createSimulationJobResponse_iamRole :: Lens' CreateSimulationJobResponse (Maybe Text)

The IAM role that allows the simulation job to call the AWS APIs that are specified in its associated policies on your behalf.

createSimulationJobResponse_maxJobDurationInSeconds :: Lens' CreateSimulationJobResponse (Maybe Integer)

The maximum simulation job duration in seconds.

createSimulationJobResponse_tags :: Lens' CreateSimulationJobResponse (Maybe (HashMap Text Text))

The list of all tags added to the simulation job.

createSimulationJobResponse_httpStatus :: Lens' CreateSimulationJobResponse Int

The response's http status code.