Amazonka.RobOMaker.DeleteWorldTemplate

Description

Deletes a world template.

Creating a Request

data DeleteWorldTemplate

See: newDeleteWorldTemplate smart constructor.

Constructors

DeleteWorldTemplate'

Fields

template :: Text

The Amazon Resource Name (arn) of the world template you want to delete.

newDeleteWorldTemplate

Arguments

:: Text
$sel:template:DeleteWorldTemplate'

-> DeleteWorldTemplate

Create a value of DeleteWorldTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:template:DeleteWorldTemplate', deleteWorldTemplate_template - The Amazon Resource Name (arn) of the world template you want to delete.

Request Lenses

deleteWorldTemplate_template :: Lens' DeleteWorldTemplate Text

The Amazon Resource Name (arn) of the world template you want to delete.

Destructuring the Response

data DeleteWorldTemplateResponse

See: newDeleteWorldTemplateResponse smart constructor.

Constructors

DeleteWorldTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteWorldTemplateResponse

Arguments

:: Int
$sel:httpStatus:DeleteWorldTemplateResponse'

-> DeleteWorldTemplateResponse

Create a value of DeleteWorldTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteWorldTemplateResponse', deleteWorldTemplateResponse_httpStatus - The response's http status code.

Response Lenses

deleteWorldTemplateResponse_httpStatus :: Lens' DeleteWorldTemplateResponse Int

The response's http status code.