Amazonka.RobOMaker.DescribeDeploymentJob

Description

Describes a deployment job.

Creating a Request

data DescribeDeploymentJob

See: newDescribeDeploymentJob smart constructor.

Constructors

DescribeDeploymentJob'

Fields

job :: Text

The Amazon Resource Name (ARN) of the deployment job.

newDescribeDeploymentJob

Arguments

:: Text
$sel:job:DescribeDeploymentJob'

-> DescribeDeploymentJob

Create a value of DescribeDeploymentJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:job:DescribeDeploymentJob', describeDeploymentJob_job - The Amazon Resource Name (ARN) of the deployment job.

Request Lenses

describeDeploymentJob_job :: Lens' DescribeDeploymentJob Text

The Amazon Resource Name (ARN) of the deployment job.

Destructuring the Response

data DescribeDeploymentJobResponse

See: newDescribeDeploymentJobResponse smart constructor.

Constructors

DescribeDeploymentJobResponse'

Fields

failureReason :: Maybe Text

A short description of the reason why the deployment job failed.

status :: Maybe DeploymentStatus

The status of the deployment job.

deploymentApplicationConfigs :: Maybe (NonEmpty DeploymentApplicationConfig)

The deployment application configuration.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the deployment job.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the deployment job was created.

failureCode :: Maybe DeploymentJobErrorCode

The deployment job failure code.

deploymentConfig :: Maybe DeploymentConfig

The deployment configuration.

fleet :: Maybe Text

The Amazon Resource Name (ARN) of the fleet.

robotDeploymentSummary :: Maybe [RobotDeployment]

A list of robot deployment summaries.

tags :: Maybe (HashMap Text Text)

The list of all tags added to the specified deployment job.

httpStatus :: Int

The response's http status code.

newDescribeDeploymentJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeDeploymentJobResponse'

-> DescribeDeploymentJobResponse

Create a value of DescribeDeploymentJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DescribeDeploymentJobResponse', describeDeploymentJobResponse_failureReason - A short description of the reason why the deployment job failed.

$sel:status:DescribeDeploymentJobResponse', describeDeploymentJobResponse_status - The status of the deployment job.

$sel:deploymentApplicationConfigs:DescribeDeploymentJobResponse', describeDeploymentJobResponse_deploymentApplicationConfigs - The deployment application configuration.

$sel:arn:DescribeDeploymentJobResponse', describeDeploymentJobResponse_arn - The Amazon Resource Name (ARN) of the deployment job.

$sel:createdAt:DescribeDeploymentJobResponse', describeDeploymentJobResponse_createdAt - The time, in milliseconds since the epoch, when the deployment job was created.

$sel:failureCode:DescribeDeploymentJobResponse', describeDeploymentJobResponse_failureCode - The deployment job failure code.

$sel:deploymentConfig:DescribeDeploymentJobResponse', describeDeploymentJobResponse_deploymentConfig - The deployment configuration.

$sel:fleet:DescribeDeploymentJobResponse', describeDeploymentJobResponse_fleet - The Amazon Resource Name (ARN) of the fleet.

$sel:robotDeploymentSummary:DescribeDeploymentJobResponse', describeDeploymentJobResponse_robotDeploymentSummary - A list of robot deployment summaries.

$sel:tags:DescribeDeploymentJobResponse', describeDeploymentJobResponse_tags - The list of all tags added to the specified deployment job.

$sel:httpStatus:DescribeDeploymentJobResponse', describeDeploymentJobResponse_httpStatus - The response's http status code.

Response Lenses

describeDeploymentJobResponse_failureReason :: Lens' DescribeDeploymentJobResponse (Maybe Text)

A short description of the reason why the deployment job failed.

describeDeploymentJobResponse_status :: Lens' DescribeDeploymentJobResponse (Maybe DeploymentStatus)

The status of the deployment job.

describeDeploymentJobResponse_deploymentApplicationConfigs :: Lens' DescribeDeploymentJobResponse (Maybe (NonEmpty DeploymentApplicationConfig))

The deployment application configuration.

describeDeploymentJobResponse_arn :: Lens' DescribeDeploymentJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the deployment job.

describeDeploymentJobResponse_createdAt :: Lens' DescribeDeploymentJobResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the deployment job was created.

describeDeploymentJobResponse_failureCode :: Lens' DescribeDeploymentJobResponse (Maybe DeploymentJobErrorCode)

The deployment job failure code.

describeDeploymentJobResponse_deploymentConfig :: Lens' DescribeDeploymentJobResponse (Maybe DeploymentConfig)

The deployment configuration.

describeDeploymentJobResponse_fleet :: Lens' DescribeDeploymentJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the fleet.

describeDeploymentJobResponse_robotDeploymentSummary :: Lens' DescribeDeploymentJobResponse (Maybe [RobotDeployment])

A list of robot deployment summaries.

describeDeploymentJobResponse_tags :: Lens' DescribeDeploymentJobResponse (Maybe (HashMap Text Text))

The list of all tags added to the specified deployment job.

describeDeploymentJobResponse_httpStatus :: Lens' DescribeDeploymentJobResponse Int

The response's http status code.