Amazonka.RobOMaker.DescribeFleet

Description

Describes a fleet.

Creating a Request

data DescribeFleet

See: newDescribeFleet smart constructor.

Constructors

DescribeFleet'

Fields

fleet :: Text

The Amazon Resource Name (ARN) of the fleet.

newDescribeFleet

Arguments

:: Text
$sel:fleet:DescribeFleet'

-> DescribeFleet

Create a value of DescribeFleet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleet:DescribeFleet', describeFleet_fleet - The Amazon Resource Name (ARN) of the fleet.

Request Lenses

describeFleet_fleet :: Lens' DescribeFleet Text

The Amazon Resource Name (ARN) of the fleet.

Destructuring the Response

data DescribeFleetResponse

See: newDescribeFleetResponse smart constructor.

Constructors

DescribeFleetResponse'

Fields

lastDeploymentJob :: Maybe Text

The Amazon Resource Name (ARN) of the last deployment job.

lastDeploymentStatus :: Maybe DeploymentStatus

The status of the last deployment.

robots :: Maybe [Robot]

A list of robots.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the fleet.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the fleet was created.

name :: Maybe Text

The name of the fleet.

lastDeploymentTime :: Maybe POSIX

The time of the last deployment.

tags :: Maybe (HashMap Text Text)

The list of all tags added to the specified fleet.

httpStatus :: Int

The response's http status code.

newDescribeFleetResponse

Arguments

:: Int
$sel:httpStatus:DescribeFleetResponse'

-> DescribeFleetResponse

Create a value of DescribeFleetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastDeploymentJob:DescribeFleetResponse', describeFleetResponse_lastDeploymentJob - The Amazon Resource Name (ARN) of the last deployment job.

$sel:lastDeploymentStatus:DescribeFleetResponse', describeFleetResponse_lastDeploymentStatus - The status of the last deployment.

$sel:robots:DescribeFleetResponse', describeFleetResponse_robots - A list of robots.

$sel:arn:DescribeFleetResponse', describeFleetResponse_arn - The Amazon Resource Name (ARN) of the fleet.

$sel:createdAt:DescribeFleetResponse', describeFleetResponse_createdAt - The time, in milliseconds since the epoch, when the fleet was created.

$sel:name:DescribeFleetResponse', describeFleetResponse_name - The name of the fleet.

$sel:lastDeploymentTime:DescribeFleetResponse', describeFleetResponse_lastDeploymentTime - The time of the last deployment.

$sel:tags:DescribeFleetResponse', describeFleetResponse_tags - The list of all tags added to the specified fleet.

$sel:httpStatus:DescribeFleetResponse', describeFleetResponse_httpStatus - The response's http status code.

Response Lenses

describeFleetResponse_lastDeploymentJob :: Lens' DescribeFleetResponse (Maybe Text)

The Amazon Resource Name (ARN) of the last deployment job.

describeFleetResponse_lastDeploymentStatus :: Lens' DescribeFleetResponse (Maybe DeploymentStatus)

The status of the last deployment.

describeFleetResponse_robots :: Lens' DescribeFleetResponse (Maybe [Robot])

A list of robots.

describeFleetResponse_arn :: Lens' DescribeFleetResponse (Maybe Text)

The Amazon Resource Name (ARN) of the fleet.

describeFleetResponse_createdAt :: Lens' DescribeFleetResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the fleet was created.

describeFleetResponse_name :: Lens' DescribeFleetResponse (Maybe Text)

The name of the fleet.

describeFleetResponse_lastDeploymentTime :: Lens' DescribeFleetResponse (Maybe UTCTime)

The time of the last deployment.

describeFleetResponse_tags :: Lens' DescribeFleetResponse (Maybe (HashMap Text Text))

The list of all tags added to the specified fleet.

describeFleetResponse_httpStatus :: Lens' DescribeFleetResponse Int

The response's http status code.