Amazonka.RobOMaker.DescribeSimulationJobBatch

Description

Describes a simulation job batch.

Creating a Request

data DescribeSimulationJobBatch

See: newDescribeSimulationJobBatch smart constructor.

Constructors

DescribeSimulationJobBatch'

Fields

batch :: Text

The id of the batch to describe.

newDescribeSimulationJobBatch

Arguments

:: Text
$sel:batch:DescribeSimulationJobBatch'

-> DescribeSimulationJobBatch

Create a value of DescribeSimulationJobBatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:batch:DescribeSimulationJobBatch', describeSimulationJobBatch_batch - The id of the batch to describe.

Request Lenses

describeSimulationJobBatch_batch :: Lens' DescribeSimulationJobBatch Text

The id of the batch to describe.

Destructuring the Response

data DescribeSimulationJobBatchResponse

See: newDescribeSimulationJobBatchResponse smart constructor.

Constructors

DescribeSimulationJobBatchResponse'

Fields

failureReason :: Maybe Text

The reason the simulation job batch failed.

status :: Maybe SimulationJobBatchStatus

The status of the batch.

Pending

The simulation job batch request is pending.

InProgress

The simulation job batch is in progress.

Failed

The simulation job batch failed. One or more simulation job requests could not be completed due to an internal failure (like InternalServiceError). See failureCode and failureReason for more information.

Completed

The simulation batch job completed. A batch is complete when (1) there are no pending simulation job requests in the batch and none of the failed simulation job requests are due to InternalServiceError and (2) when all created simulation jobs have reached a terminal state (for example, Completed or Failed).

Canceled

The simulation batch job was cancelled.

Canceling

The simulation batch job is being cancelled.

Completing

The simulation batch job is completing.

TimingOut

The simulation job batch is timing out.
If a batch timing out, and there are pending requests that were failing due to an internal failure (like InternalServiceError), the batch status will be Failed. If there are no such failing request, the batch status will be TimedOut.

TimedOut

The simulation batch job timed out.

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation job batch was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the batch.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation job batch was created.

failureCode :: Maybe SimulationJobBatchErrorCode

The failure code of the simulation job batch.

failedRequests :: Maybe [FailedCreateSimulationJobRequest]

A list of failed create simulation job requests. The request failed to be created into a simulation job. Failed requests do not have a simulation job ID.

batchPolicy :: Maybe BatchPolicy

The batch policy.

createdRequests :: Maybe [SimulationJobSummary]

A list of created simulation job summaries.

pendingRequests :: Maybe (NonEmpty SimulationJobRequest)

A list of pending simulation job requests. These requests have not yet been created into simulation jobs.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

tags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the simulation job batch.

httpStatus :: Int

The response's http status code.

newDescribeSimulationJobBatchResponse

Arguments

:: Int
$sel:httpStatus:DescribeSimulationJobBatchResponse'

-> DescribeSimulationJobBatchResponse

Create a value of DescribeSimulationJobBatchResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_failureReason - The reason the simulation job batch failed.

$sel:status:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_status - The status of the batch.

Pending

The simulation job batch request is pending.

InProgress

The simulation job batch is in progress.

Failed

The simulation job batch failed. One or more simulation job requests could not be completed due to an internal failure (like InternalServiceError). See failureCode and failureReason for more information.

Completed

The simulation batch job completed. A batch is complete when (1) there are no pending simulation job requests in the batch and none of the failed simulation job requests are due to InternalServiceError and (2) when all created simulation jobs have reached a terminal state (for example, Completed or Failed).

Canceled

The simulation batch job was cancelled.

Canceling

The simulation batch job is being cancelled.

Completing

The simulation batch job is completing.

TimingOut

The simulation job batch is timing out.
If a batch timing out, and there are pending requests that were failing due to an internal failure (like InternalServiceError), the batch status will be Failed. If there are no such failing request, the batch status will be TimedOut.

TimedOut

The simulation batch job timed out.

$sel:lastUpdatedAt:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_lastUpdatedAt - The time, in milliseconds since the epoch, when the simulation job batch was last updated.

$sel:arn:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_arn - The Amazon Resource Name (ARN) of the batch.

$sel:createdAt:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_createdAt - The time, in milliseconds since the epoch, when the simulation job batch was created.

$sel:failureCode:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_failureCode - The failure code of the simulation job batch.

$sel:failedRequests:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_failedRequests - A list of failed create simulation job requests. The request failed to be created into a simulation job. Failed requests do not have a simulation job ID.

$sel:batchPolicy:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_batchPolicy - The batch policy.

$sel:createdRequests:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_createdRequests - A list of created simulation job summaries.

$sel:pendingRequests:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_pendingRequests - A list of pending simulation job requests. These requests have not yet been created into simulation jobs.

$sel:clientRequestToken:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:tags:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_tags - A map that contains tag keys and tag values that are attached to the simulation job batch.

$sel:httpStatus:DescribeSimulationJobBatchResponse', describeSimulationJobBatchResponse_httpStatus - The response's http status code.

Response Lenses

describeSimulationJobBatchResponse_failureReason :: Lens' DescribeSimulationJobBatchResponse (Maybe Text)

The reason the simulation job batch failed.

describeSimulationJobBatchResponse_status :: Lens' DescribeSimulationJobBatchResponse (Maybe SimulationJobBatchStatus)

The status of the batch.

Pending

The simulation job batch request is pending.

InProgress

The simulation job batch is in progress.

Failed

The simulation job batch failed. One or more simulation job requests could not be completed due to an internal failure (like InternalServiceError). See failureCode and failureReason for more information.

Completed

The simulation batch job completed. A batch is complete when (1) there are no pending simulation job requests in the batch and none of the failed simulation job requests are due to InternalServiceError and (2) when all created simulation jobs have reached a terminal state (for example, Completed or Failed).

Canceled

The simulation batch job was cancelled.

Canceling

The simulation batch job is being cancelled.

Completing

The simulation batch job is completing.

TimingOut

The simulation job batch is timing out.
If a batch timing out, and there are pending requests that were failing due to an internal failure (like InternalServiceError), the batch status will be Failed. If there are no such failing request, the batch status will be TimedOut.

TimedOut

The simulation batch job timed out.

describeSimulationJobBatchResponse_lastUpdatedAt :: Lens' DescribeSimulationJobBatchResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation job batch was last updated.

describeSimulationJobBatchResponse_arn :: Lens' DescribeSimulationJobBatchResponse (Maybe Text)

The Amazon Resource Name (ARN) of the batch.

describeSimulationJobBatchResponse_createdAt :: Lens' DescribeSimulationJobBatchResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation job batch was created.

describeSimulationJobBatchResponse_failureCode :: Lens' DescribeSimulationJobBatchResponse (Maybe SimulationJobBatchErrorCode)

The failure code of the simulation job batch.

describeSimulationJobBatchResponse_failedRequests :: Lens' DescribeSimulationJobBatchResponse (Maybe [FailedCreateSimulationJobRequest])

A list of failed create simulation job requests. The request failed to be created into a simulation job. Failed requests do not have a simulation job ID.

describeSimulationJobBatchResponse_batchPolicy :: Lens' DescribeSimulationJobBatchResponse (Maybe BatchPolicy)

The batch policy.

describeSimulationJobBatchResponse_createdRequests :: Lens' DescribeSimulationJobBatchResponse (Maybe [SimulationJobSummary])

A list of created simulation job summaries.

describeSimulationJobBatchResponse_pendingRequests :: Lens' DescribeSimulationJobBatchResponse (Maybe (NonEmpty SimulationJobRequest))

A list of pending simulation job requests. These requests have not yet been created into simulation jobs.

describeSimulationJobBatchResponse_clientRequestToken :: Lens' DescribeSimulationJobBatchResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

describeSimulationJobBatchResponse_tags :: Lens' DescribeSimulationJobBatchResponse (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the simulation job batch.

describeSimulationJobBatchResponse_httpStatus :: Lens' DescribeSimulationJobBatchResponse Int

The response's http status code.