Amazonka.RobOMaker.ListSimulationJobBatches

Description

Returns a list simulation job batches. You can optionally provide filters to retrieve specific simulation batch jobs.

This operation returns paginated results.

Creating a Request

data ListSimulationJobBatches

See: newListSimulationJobBatches smart constructor.

Constructors

ListSimulationJobBatches'

Fields

filters :: Maybe (NonEmpty Filter)

Optional filters to limit results.

nextToken :: Maybe Text

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListSimulationJobBatches again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

maxResults :: Maybe Int

When this parameter is used, ListSimulationJobBatches only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListSimulationJobBatches request with the returned nextToken value.

newListSimulationJobBatches :: ListSimulationJobBatches

Create a value of ListSimulationJobBatches with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListSimulationJobBatches', listSimulationJobBatches_filters - Optional filters to limit results.

$sel:nextToken:ListSimulationJobBatches', listSimulationJobBatches_nextToken - If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListSimulationJobBatches again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:maxResults:ListSimulationJobBatches', listSimulationJobBatches_maxResults - When this parameter is used, ListSimulationJobBatches only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListSimulationJobBatches request with the returned nextToken value.

Request Lenses

listSimulationJobBatches_filters :: Lens' ListSimulationJobBatches (Maybe (NonEmpty Filter))

Optional filters to limit results.

listSimulationJobBatches_nextToken :: Lens' ListSimulationJobBatches (Maybe Text)

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListSimulationJobBatches again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listSimulationJobBatches_maxResults :: Lens' ListSimulationJobBatches (Maybe Int)

When this parameter is used, ListSimulationJobBatches only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListSimulationJobBatches request with the returned nextToken value.

Destructuring the Response

data ListSimulationJobBatchesResponse

See: newListSimulationJobBatchesResponse smart constructor.

Constructors

ListSimulationJobBatchesResponse'

Fields

nextToken :: Maybe Text

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListSimulationJobBatches again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

simulationJobBatchSummaries :: Maybe [SimulationJobBatchSummary]

A list of simulation job batch summaries.

httpStatus :: Int

The response's http status code.

newListSimulationJobBatchesResponse

Arguments

:: Int
$sel:httpStatus:ListSimulationJobBatchesResponse'

-> ListSimulationJobBatchesResponse

Create a value of ListSimulationJobBatchesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSimulationJobBatches', listSimulationJobBatchesResponse_nextToken - If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListSimulationJobBatches again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:simulationJobBatchSummaries:ListSimulationJobBatchesResponse', listSimulationJobBatchesResponse_simulationJobBatchSummaries - A list of simulation job batch summaries.

$sel:httpStatus:ListSimulationJobBatchesResponse', listSimulationJobBatchesResponse_httpStatus - The response's http status code.

Response Lenses

listSimulationJobBatchesResponse_nextToken :: Lens' ListSimulationJobBatchesResponse (Maybe Text)

If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListSimulationJobBatches again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listSimulationJobBatchesResponse_simulationJobBatchSummaries :: Lens' ListSimulationJobBatchesResponse (Maybe [SimulationJobBatchSummary])

A list of simulation job batch summaries.

listSimulationJobBatchesResponse_httpStatus :: Lens' ListSimulationJobBatchesResponse Int

The response's http status code.