Amazonka.RobOMaker.StartSimulationJobBatch

Description

Starts a new simulation job batch. The batch is defined using one or more SimulationJobRequest objects.

Creating a Request

data StartSimulationJobBatch

See: newStartSimulationJobBatch smart constructor.

Constructors

StartSimulationJobBatch'

Fields

batchPolicy :: Maybe BatchPolicy

The batch policy.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

tags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the deployment job batch.

createSimulationJobRequests :: NonEmpty SimulationJobRequest

A list of simulation job requests to create in the batch.

newStartSimulationJobBatch

Arguments

:: NonEmpty SimulationJobRequest
$sel:createSimulationJobRequests:StartSimulationJobBatch'

-> StartSimulationJobBatch

Create a value of StartSimulationJobBatch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:batchPolicy:StartSimulationJobBatch', startSimulationJobBatch_batchPolicy - The batch policy.

$sel:clientRequestToken:StartSimulationJobBatch', startSimulationJobBatch_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:tags:StartSimulationJobBatch', startSimulationJobBatch_tags - A map that contains tag keys and tag values that are attached to the deployment job batch.

$sel:createSimulationJobRequests:StartSimulationJobBatch', startSimulationJobBatch_createSimulationJobRequests - A list of simulation job requests to create in the batch.

Request Lenses

startSimulationJobBatch_batchPolicy :: Lens' StartSimulationJobBatch (Maybe BatchPolicy)

The batch policy.

startSimulationJobBatch_clientRequestToken :: Lens' StartSimulationJobBatch (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

startSimulationJobBatch_tags :: Lens' StartSimulationJobBatch (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the deployment job batch.

startSimulationJobBatch_createSimulationJobRequests :: Lens' StartSimulationJobBatch (NonEmpty SimulationJobRequest)

A list of simulation job requests to create in the batch.

Destructuring the Response

data StartSimulationJobBatchResponse

See: newStartSimulationJobBatchResponse smart constructor.

Constructors

StartSimulationJobBatchResponse'

Fields

failureReason :: Maybe Text

The reason the simulation job batch failed.

status :: Maybe SimulationJobBatchStatus

The status of the simulation job batch.

Pending

The simulation job batch request is pending.

InProgress

The simulation job batch is in progress.

Failed

The simulation job batch failed. One or more simulation job requests could not be completed due to an internal failure (like InternalServiceError). See failureCode and failureReason for more information.

Completed

The simulation batch job completed. A batch is complete when (1) there are no pending simulation job requests in the batch and none of the failed simulation job requests are due to InternalServiceError and (2) when all created simulation jobs have reached a terminal state (for example, Completed or Failed).

Canceled

The simulation batch job was cancelled.

Canceling

The simulation batch job is being cancelled.

Completing

The simulation batch job is completing.

TimingOut

The simulation job batch is timing out.
If a batch timing out, and there are pending requests that were failing due to an internal failure (like InternalServiceError), the batch status will be Failed. If there are no such failing request, the batch status will be TimedOut.

TimedOut

The simulation batch job timed out.

arn :: Maybe Text

The Amazon Resource Name (arn) of the batch.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the simulation job batch was created.

failureCode :: Maybe SimulationJobBatchErrorCode

The failure code if the simulation job batch failed.

failedRequests :: Maybe [FailedCreateSimulationJobRequest]

A list of failed simulation job requests. The request failed to be created into a simulation job. Failed requests do not have a simulation job ID.

batchPolicy :: Maybe BatchPolicy

The batch policy.

createdRequests :: Maybe [SimulationJobSummary]

A list of created simulation job request summaries.

pendingRequests :: Maybe (NonEmpty SimulationJobRequest)

A list of pending simulation job requests. These requests have not yet been created into simulation jobs.

clientRequestToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

tags :: Maybe (HashMap Text Text)

A map that contains tag keys and tag values that are attached to the deployment job batch.

httpStatus :: Int

The response's http status code.

newStartSimulationJobBatchResponse

Arguments

:: Int
$sel:httpStatus:StartSimulationJobBatchResponse'

-> StartSimulationJobBatchResponse

Create a value of StartSimulationJobBatchResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:StartSimulationJobBatchResponse', startSimulationJobBatchResponse_failureReason - The reason the simulation job batch failed.

$sel:status:StartSimulationJobBatchResponse', startSimulationJobBatchResponse_status - The status of the simulation job batch.

Pending

The simulation job batch request is pending.

InProgress

The simulation job batch is in progress.

Failed

The simulation job batch failed. One or more simulation job requests could not be completed due to an internal failure (like InternalServiceError). See failureCode and failureReason for more information.

Completed

The simulation batch job completed. A batch is complete when (1) there are no pending simulation job requests in the batch and none of the failed simulation job requests are due to InternalServiceError and (2) when all created simulation jobs have reached a terminal state (for example, Completed or Failed).

Canceled

The simulation batch job was cancelled.

Canceling

The simulation batch job is being cancelled.

Completing

The simulation batch job is completing.

TimingOut

The simulation job batch is timing out.
If a batch timing out, and there are pending requests that were failing due to an internal failure (like InternalServiceError), the batch status will be Failed. If there are no such failing request, the batch status will be TimedOut.

TimedOut

The simulation batch job timed out.

$sel:arn:StartSimulationJobBatchResponse', startSimulationJobBatchResponse_arn - The Amazon Resource Name (arn) of the batch.

$sel:createdAt:StartSimulationJobBatchResponse', startSimulationJobBatchResponse_createdAt - The time, in milliseconds since the epoch, when the simulation job batch was created.

$sel:failureCode:StartSimulationJobBatchResponse', startSimulationJobBatchResponse_failureCode - The failure code if the simulation job batch failed.

$sel:failedRequests:StartSimulationJobBatchResponse', startSimulationJobBatchResponse_failedRequests - A list of failed simulation job requests. The request failed to be created into a simulation job. Failed requests do not have a simulation job ID.

$sel:batchPolicy:StartSimulationJobBatch', startSimulationJobBatchResponse_batchPolicy - The batch policy.

$sel:createdRequests:StartSimulationJobBatchResponse', startSimulationJobBatchResponse_createdRequests - A list of created simulation job request summaries.

$sel:pendingRequests:StartSimulationJobBatchResponse', startSimulationJobBatchResponse_pendingRequests - A list of pending simulation job requests. These requests have not yet been created into simulation jobs.

$sel:clientRequestToken:StartSimulationJobBatch', startSimulationJobBatchResponse_clientRequestToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:tags:StartSimulationJobBatch', startSimulationJobBatchResponse_tags - A map that contains tag keys and tag values that are attached to the deployment job batch.

$sel:httpStatus:StartSimulationJobBatchResponse', startSimulationJobBatchResponse_httpStatus - The response's http status code.

Response Lenses

startSimulationJobBatchResponse_failureReason :: Lens' StartSimulationJobBatchResponse (Maybe Text)

The reason the simulation job batch failed.

startSimulationJobBatchResponse_status :: Lens' StartSimulationJobBatchResponse (Maybe SimulationJobBatchStatus)

The status of the simulation job batch.

Pending

The simulation job batch request is pending.

InProgress

The simulation job batch is in progress.

Failed

The simulation job batch failed. One or more simulation job requests could not be completed due to an internal failure (like InternalServiceError). See failureCode and failureReason for more information.

Completed

The simulation batch job completed. A batch is complete when (1) there are no pending simulation job requests in the batch and none of the failed simulation job requests are due to InternalServiceError and (2) when all created simulation jobs have reached a terminal state (for example, Completed or Failed).

Canceled

The simulation batch job was cancelled.

Canceling

The simulation batch job is being cancelled.

Completing

The simulation batch job is completing.

TimingOut

The simulation job batch is timing out.
If a batch timing out, and there are pending requests that were failing due to an internal failure (like InternalServiceError), the batch status will be Failed. If there are no such failing request, the batch status will be TimedOut.

TimedOut

The simulation batch job timed out.

startSimulationJobBatchResponse_arn :: Lens' StartSimulationJobBatchResponse (Maybe Text)

The Amazon Resource Name (arn) of the batch.

startSimulationJobBatchResponse_createdAt :: Lens' StartSimulationJobBatchResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the simulation job batch was created.

startSimulationJobBatchResponse_failureCode :: Lens' StartSimulationJobBatchResponse (Maybe SimulationJobBatchErrorCode)

The failure code if the simulation job batch failed.

startSimulationJobBatchResponse_failedRequests :: Lens' StartSimulationJobBatchResponse (Maybe [FailedCreateSimulationJobRequest])

A list of failed simulation job requests. The request failed to be created into a simulation job. Failed requests do not have a simulation job ID.

startSimulationJobBatchResponse_batchPolicy :: Lens' StartSimulationJobBatchResponse (Maybe BatchPolicy)

The batch policy.

startSimulationJobBatchResponse_createdRequests :: Lens' StartSimulationJobBatchResponse (Maybe [SimulationJobSummary])

A list of created simulation job request summaries.

startSimulationJobBatchResponse_pendingRequests :: Lens' StartSimulationJobBatchResponse (Maybe (NonEmpty SimulationJobRequest))

A list of pending simulation job requests. These requests have not yet been created into simulation jobs.

startSimulationJobBatchResponse_clientRequestToken :: Lens' StartSimulationJobBatchResponse (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

startSimulationJobBatchResponse_tags :: Lens' StartSimulationJobBatchResponse (Maybe (HashMap Text Text))

A map that contains tag keys and tag values that are attached to the deployment job batch.

startSimulationJobBatchResponse_httpStatus :: Lens' StartSimulationJobBatchResponse Int

The response's http status code.