Amazonka.RobOMaker.Types.DataSource

Description

Documentation

data DataSource

Information about a data source.

See: newDataSource smart constructor.

Constructors

DataSource'

Fields

s3Keys :: Maybe [S3KeyOutput]

The list of S3 keys identifying the data source files.

destination :: Maybe Text

The location where your files are mounted in the container image.

If you've specified the type of the data source as an Archive, you must provide an Amazon S3 object key to your archive. The object key must point to either a .zip or .tar.gz file.

If you've specified the type of the data source as a Prefix, you provide the Amazon S3 prefix that points to the files that you are using for your data source.

If you've specified the type of the data source as a File, you provide the Amazon S3 path to the file that you're using as your data source.

name :: Maybe Text

The name of the data source.

type' :: Maybe DataSourceType

The data type for the data source that you're using for your container image or simulation job. You can use this field to specify whether your data source is an Archive, an Amazon S3 prefix, or a file.

If you don't specify a field, the default value is File.

s3Bucket :: Maybe Text

The S3 bucket where the data files are located.

newDataSource :: DataSource

Create a value of DataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Keys:DataSource', dataSource_s3Keys - The list of S3 keys identifying the data source files.

$sel:destination:DataSource', dataSource_destination - The location where your files are mounted in the container image.

If you've specified the type of the data source as an Archive, you must provide an Amazon S3 object key to your archive. The object key must point to either a .zip or .tar.gz file.

If you've specified the type of the data source as a Prefix, you provide the Amazon S3 prefix that points to the files that you are using for your data source.

If you've specified the type of the data source as a File, you provide the Amazon S3 path to the file that you're using as your data source.

$sel:name:DataSource', dataSource_name - The name of the data source.

$sel:type':DataSource', dataSource_type - The data type for the data source that you're using for your container image or simulation job. You can use this field to specify whether your data source is an Archive, an Amazon S3 prefix, or a file.

If you don't specify a field, the default value is File.

$sel:s3Bucket:DataSource', dataSource_s3Bucket - The S3 bucket where the data files are located.

dataSource_s3Keys :: Lens' DataSource (Maybe [S3KeyOutput])

The list of S3 keys identifying the data source files.

dataSource_destination :: Lens' DataSource (Maybe Text)

The location where your files are mounted in the container image.

If you've specified the type of the data source as an Archive, you must provide an Amazon S3 object key to your archive. The object key must point to either a .zip or .tar.gz file.

If you've specified the type of the data source as a Prefix, you provide the Amazon S3 prefix that points to the files that you are using for your data source.

If you've specified the type of the data source as a File, you provide the Amazon S3 path to the file that you're using as your data source.

dataSource_name :: Lens' DataSource (Maybe Text)

The name of the data source.

dataSource_type :: Lens' DataSource (Maybe DataSourceType)

The data type for the data source that you're using for your container image or simulation job. You can use this field to specify whether your data source is an Archive, an Amazon S3 prefix, or a file.

If you don't specify a field, the default value is File.

dataSource_s3Bucket :: Lens' DataSource (Maybe Text)

The S3 bucket where the data files are located.