Amazonka.RobOMaker.Types.DeploymentLaunchConfig

Description

Documentation

data DeploymentLaunchConfig

Configuration information for a deployment launch.

See: newDeploymentLaunchConfig smart constructor.

Constructors

DeploymentLaunchConfig'

Fields

preLaunchFile :: Maybe Text

The deployment pre-launch file. This file will be executed prior to the launch file.

postLaunchFile :: Maybe Text

The deployment post-launch file. This file will be executed after the launch file.

environmentVariables :: Maybe (HashMap Text Text)

An array of key/value pairs specifying environment variables for the robot application

packageName :: Text

The package name.

launchFile :: Text

The launch file name.

newDeploymentLaunchConfig

Arguments

:: Text
$sel:packageName:DeploymentLaunchConfig'

-> Text
$sel:launchFile:DeploymentLaunchConfig'

-> DeploymentLaunchConfig

Create a value of DeploymentLaunchConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preLaunchFile:DeploymentLaunchConfig', deploymentLaunchConfig_preLaunchFile - The deployment pre-launch file. This file will be executed prior to the launch file.

$sel:postLaunchFile:DeploymentLaunchConfig', deploymentLaunchConfig_postLaunchFile - The deployment post-launch file. This file will be executed after the launch file.

$sel:environmentVariables:DeploymentLaunchConfig', deploymentLaunchConfig_environmentVariables - An array of key/value pairs specifying environment variables for the robot application

$sel:packageName:DeploymentLaunchConfig', deploymentLaunchConfig_packageName - The package name.

$sel:launchFile:DeploymentLaunchConfig', deploymentLaunchConfig_launchFile - The launch file name.

deploymentLaunchConfig_preLaunchFile :: Lens' DeploymentLaunchConfig (Maybe Text)

The deployment pre-launch file. This file will be executed prior to the launch file.

deploymentLaunchConfig_postLaunchFile :: Lens' DeploymentLaunchConfig (Maybe Text)

The deployment post-launch file. This file will be executed after the launch file.

deploymentLaunchConfig_environmentVariables :: Lens' DeploymentLaunchConfig (Maybe (HashMap Text Text))

An array of key/value pairs specifying environment variables for the robot application

deploymentLaunchConfig_packageName :: Lens' DeploymentLaunchConfig Text

The package name.

deploymentLaunchConfig_launchFile :: Lens' DeploymentLaunchConfig Text

The launch file name.