Amazonka.RobOMaker.Types.LaunchConfig

Description

Documentation

data LaunchConfig

Information about a launch configuration.

See: newLaunchConfig smart constructor.

Constructors

LaunchConfig'

Fields

command :: Maybe [Text]

If you've specified General as the value for your RobotSoftwareSuite, you can use this field to specify a list of commands for your container image.

If you've specified SimulationRuntime as the value for your SimulationSoftwareSuite, you can use this field to specify a list of commands for your container image.

packageName :: Maybe Text

The package name.

portForwardingConfig :: Maybe PortForwardingConfig

The port forwarding configuration.

launchFile :: Maybe Text

The launch file name.

environmentVariables :: Maybe (HashMap Text Text)

The environment variables for the application launch.

streamUI :: Maybe Bool

Boolean indicating whether a streaming session will be configured for the application. If True, AWS RoboMaker will configure a connection so you can interact with your application as it is running in the simulation. You must configure and launch the component. It must have a graphical user interface.

newLaunchConfig :: LaunchConfig

Create a value of LaunchConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:command:LaunchConfig', launchConfig_command - If you've specified General as the value for your RobotSoftwareSuite, you can use this field to specify a list of commands for your container image.

If you've specified SimulationRuntime as the value for your SimulationSoftwareSuite, you can use this field to specify a list of commands for your container image.

$sel:packageName:LaunchConfig', launchConfig_packageName - The package name.

$sel:portForwardingConfig:LaunchConfig', launchConfig_portForwardingConfig - The port forwarding configuration.

$sel:launchFile:LaunchConfig', launchConfig_launchFile - The launch file name.

$sel:environmentVariables:LaunchConfig', launchConfig_environmentVariables - The environment variables for the application launch.

$sel:streamUI:LaunchConfig', launchConfig_streamUI - Boolean indicating whether a streaming session will be configured for the application. If True, AWS RoboMaker will configure a connection so you can interact with your application as it is running in the simulation. You must configure and launch the component. It must have a graphical user interface.

launchConfig_command :: Lens' LaunchConfig (Maybe [Text])

If you've specified General as the value for your RobotSoftwareSuite, you can use this field to specify a list of commands for your container image.

If you've specified SimulationRuntime as the value for your SimulationSoftwareSuite, you can use this field to specify a list of commands for your container image.

launchConfig_packageName :: Lens' LaunchConfig (Maybe Text)

The package name.

launchConfig_portForwardingConfig :: Lens' LaunchConfig (Maybe PortForwardingConfig)

The port forwarding configuration.

launchConfig_launchFile :: Lens' LaunchConfig (Maybe Text)

The launch file name.

launchConfig_environmentVariables :: Lens' LaunchConfig (Maybe (HashMap Text Text))

The environment variables for the application launch.

launchConfig_streamUI :: Lens' LaunchConfig (Maybe Bool)

Boolean indicating whether a streaming session will be configured for the application. If True, AWS RoboMaker will configure a connection so you can interact with your application as it is running in the simulation. You must configure and launch the component. It must have a graphical user interface.