Amazonka.RobOMaker.Types.ProgressDetail

Description

Documentation

data ProgressDetail

Information about the progress of a deployment job.

See: newProgressDetail smart constructor.

Constructors

ProgressDetail'

Fields

currentProgress :: Maybe RobotDeploymentStep

The current progress status.

Validating

Validating the deployment.

DownloadingExtracting

Downloading and extracting the bundle on the robot.

ExecutingPreLaunch

Executing pre-launch script(s) if provided.

Launching

Launching the robot application.

ExecutingPostLaunch

Executing post-launch script(s) if provided.

Finished

Deployment is complete.

estimatedTimeRemainingSeconds :: Maybe Int

Estimated amount of time in seconds remaining in the step. This currently only applies to the Downloading/Extracting step of the deployment. It is empty for other steps.

targetResource :: Maybe Text

The Amazon Resource Name (ARN) of the deployment job.

percentDone :: Maybe Double

Precentage of the step that is done. This currently only applies to the Downloading/Extracting step of the deployment. It is empty for other steps.

newProgressDetail :: ProgressDetail

Create a value of ProgressDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentProgress:ProgressDetail', progressDetail_currentProgress - The current progress status.

Validating

Validating the deployment.

DownloadingExtracting

Downloading and extracting the bundle on the robot.

ExecutingPreLaunch

Executing pre-launch script(s) if provided.

Launching

Launching the robot application.

ExecutingPostLaunch

Executing post-launch script(s) if provided.

Finished

Deployment is complete.

$sel:estimatedTimeRemainingSeconds:ProgressDetail', progressDetail_estimatedTimeRemainingSeconds - Estimated amount of time in seconds remaining in the step. This currently only applies to the Downloading/Extracting step of the deployment. It is empty for other steps.

$sel:targetResource:ProgressDetail', progressDetail_targetResource - The Amazon Resource Name (ARN) of the deployment job.

$sel:percentDone:ProgressDetail', progressDetail_percentDone - Precentage of the step that is done. This currently only applies to the Downloading/Extracting step of the deployment. It is empty for other steps.

progressDetail_currentProgress :: Lens' ProgressDetail (Maybe RobotDeploymentStep)

The current progress status.

Validating

Validating the deployment.

DownloadingExtracting

Downloading and extracting the bundle on the robot.

ExecutingPreLaunch

Executing pre-launch script(s) if provided.

Launching

Launching the robot application.

ExecutingPostLaunch

Executing post-launch script(s) if provided.

Finished

Deployment is complete.

progressDetail_estimatedTimeRemainingSeconds :: Lens' ProgressDetail (Maybe Int)

Estimated amount of time in seconds remaining in the step. This currently only applies to the Downloading/Extracting step of the deployment. It is empty for other steps.

progressDetail_targetResource :: Lens' ProgressDetail (Maybe Text)

The Amazon Resource Name (ARN) of the deployment job.

progressDetail_percentDone :: Lens' ProgressDetail (Maybe Double)

Precentage of the step that is done. This currently only applies to the Downloading/Extracting step of the deployment. It is empty for other steps.