Amazonka.RobOMaker.Types.RobotApplicationConfig

Description

Documentation

data RobotApplicationConfig

Application configuration information for a robot.

See: newRobotApplicationConfig smart constructor.

Constructors

RobotApplicationConfig'

Fields

useDefaultUploadConfigurations :: Maybe Bool

A Boolean indicating whether to use default upload configurations. By default, .ros and .gazebo files are uploaded when the application terminates and all ROS topics will be recorded.

If you set this value, you must specify an outputLocation.

useDefaultTools :: Maybe Bool

A Boolean indicating whether to use default robot application tools. The default tools are rviz, rqt, terminal and rosbag record. The default is False.

applicationVersion :: Maybe Text

The version of the robot application.

uploadConfigurations :: Maybe [UploadConfiguration]

The upload configurations for the robot application.

tools :: Maybe [Tool]

Information about tools configured for the robot application.

application :: Text

The application information for the robot application.

launchConfig :: LaunchConfig

The launch configuration for the robot application.

newRobotApplicationConfig

Arguments

:: Text
$sel:application:RobotApplicationConfig'

-> LaunchConfig
$sel:launchConfig:RobotApplicationConfig'

-> RobotApplicationConfig

Create a value of RobotApplicationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:useDefaultUploadConfigurations:RobotApplicationConfig', robotApplicationConfig_useDefaultUploadConfigurations - A Boolean indicating whether to use default upload configurations. By default, .ros and .gazebo files are uploaded when the application terminates and all ROS topics will be recorded.

If you set this value, you must specify an outputLocation.

$sel:useDefaultTools:RobotApplicationConfig', robotApplicationConfig_useDefaultTools - A Boolean indicating whether to use default robot application tools. The default tools are rviz, rqt, terminal and rosbag record. The default is False.

$sel:applicationVersion:RobotApplicationConfig', robotApplicationConfig_applicationVersion - The version of the robot application.

$sel:uploadConfigurations:RobotApplicationConfig', robotApplicationConfig_uploadConfigurations - The upload configurations for the robot application.

$sel:tools:RobotApplicationConfig', robotApplicationConfig_tools - Information about tools configured for the robot application.

$sel:application:RobotApplicationConfig', robotApplicationConfig_application - The application information for the robot application.

$sel:launchConfig:RobotApplicationConfig', robotApplicationConfig_launchConfig - The launch configuration for the robot application.

robotApplicationConfig_useDefaultUploadConfigurations :: Lens' RobotApplicationConfig (Maybe Bool)

A Boolean indicating whether to use default upload configurations. By default, .ros and .gazebo files are uploaded when the application terminates and all ROS topics will be recorded.

If you set this value, you must specify an outputLocation.

robotApplicationConfig_useDefaultTools :: Lens' RobotApplicationConfig (Maybe Bool)

A Boolean indicating whether to use default robot application tools. The default tools are rviz, rqt, terminal and rosbag record. The default is False.

robotApplicationConfig_applicationVersion :: Lens' RobotApplicationConfig (Maybe Text)

The version of the robot application.

robotApplicationConfig_uploadConfigurations :: Lens' RobotApplicationConfig (Maybe [UploadConfiguration])

The upload configurations for the robot application.

robotApplicationConfig_tools :: Lens' RobotApplicationConfig (Maybe [Tool])

Information about tools configured for the robot application.

robotApplicationConfig_application :: Lens' RobotApplicationConfig Text

The application information for the robot application.

robotApplicationConfig_launchConfig :: Lens' RobotApplicationConfig LaunchConfig

The launch configuration for the robot application.