Amazonka.RobOMaker.Types.Tool

Description

Documentation

data Tool

Information about a tool. Tools are used in a simulation job.

See: newTool smart constructor.

Constructors

Tool'

Fields

streamOutputToCloudWatch :: Maybe Bool

Boolean indicating whether logs will be recorded in CloudWatch for the tool. The default is False.

exitBehavior :: Maybe ExitBehavior

Exit behavior determines what happens when your tool quits running. RESTART will cause your tool to be restarted. FAIL will cause your job to exit. The default is RESTART.

streamUI :: Maybe Bool

Boolean indicating whether a streaming session will be configured for the tool. If True, AWS RoboMaker will configure a connection so you can interact with the tool as it is running in the simulation. It must have a graphical user interface. The default is False.

name :: Text

The name of the tool.

command :: Text

Command-line arguments for the tool. It must include the tool executable name.

newTool

Arguments

:: Text
$sel:name:Tool'

-> Text
$sel:command:Tool'

-> Tool

Create a value of Tool with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamOutputToCloudWatch:Tool', tool_streamOutputToCloudWatch - Boolean indicating whether logs will be recorded in CloudWatch for the tool. The default is False.

$sel:exitBehavior:Tool', tool_exitBehavior - Exit behavior determines what happens when your tool quits running. RESTART will cause your tool to be restarted. FAIL will cause your job to exit. The default is RESTART.

$sel:streamUI:Tool', tool_streamUI - Boolean indicating whether a streaming session will be configured for the tool. If True, AWS RoboMaker will configure a connection so you can interact with the tool as it is running in the simulation. It must have a graphical user interface. The default is False.

$sel:name:Tool', tool_name - The name of the tool.

$sel:command:Tool', tool_command - Command-line arguments for the tool. It must include the tool executable name.

tool_streamOutputToCloudWatch :: Lens' Tool (Maybe Bool)

Boolean indicating whether logs will be recorded in CloudWatch for the tool. The default is False.

tool_exitBehavior :: Lens' Tool (Maybe ExitBehavior)

Exit behavior determines what happens when your tool quits running. RESTART will cause your tool to be restarted. FAIL will cause your job to exit. The default is RESTART.

tool_streamUI :: Lens' Tool (Maybe Bool)

Boolean indicating whether a streaming session will be configured for the tool. If True, AWS RoboMaker will configure a connection so you can interact with the tool as it is running in the simulation. It must have a graphical user interface. The default is False.

tool_name :: Lens' Tool Text

The name of the tool.

tool_command :: Lens' Tool Text

Command-line arguments for the tool. It must include the tool executable name.