Amazonka.RobOMaker.Types.WorldGenerationJobSummary

Description

Documentation

data WorldGenerationJobSummary

Information about a world generator job.

See: newWorldGenerationJobSummary smart constructor.

Constructors

WorldGenerationJobSummary'

Fields

status :: Maybe WorldGenerationJobStatus

The status of the world generator job:

Pending

The world generator job request is pending.

Running

The world generator job is running.

Completed

The world generator job completed.

Failed

The world generator job failed. See failureCode for more information.

PartialFailed

Some worlds did not generate.

Canceled

The world generator job was cancelled.

Canceling

The world generator job is being cancelled.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the world generator job.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the world generator job was created.

worldCount :: Maybe WorldCount

Information about the world count.

succeededWorldCount :: Maybe Int

The number of worlds that were generated.

failedWorldCount :: Maybe Int

The number of worlds that failed.

template :: Maybe Text

The Amazon Resource Name (arn) of the world template.

newWorldGenerationJobSummary :: WorldGenerationJobSummary

Create a value of WorldGenerationJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:WorldGenerationJobSummary', worldGenerationJobSummary_status - The status of the world generator job:

Pending

The world generator job request is pending.

Running

The world generator job is running.

Completed

The world generator job completed.

Failed

The world generator job failed. See failureCode for more information.

PartialFailed

Some worlds did not generate.

Canceled

The world generator job was cancelled.

Canceling

The world generator job is being cancelled.

$sel:arn:WorldGenerationJobSummary', worldGenerationJobSummary_arn - The Amazon Resource Name (ARN) of the world generator job.

$sel:createdAt:WorldGenerationJobSummary', worldGenerationJobSummary_createdAt - The time, in milliseconds since the epoch, when the world generator job was created.

$sel:worldCount:WorldGenerationJobSummary', worldGenerationJobSummary_worldCount - Information about the world count.

$sel:succeededWorldCount:WorldGenerationJobSummary', worldGenerationJobSummary_succeededWorldCount - The number of worlds that were generated.

$sel:failedWorldCount:WorldGenerationJobSummary', worldGenerationJobSummary_failedWorldCount - The number of worlds that failed.

$sel:template:WorldGenerationJobSummary', worldGenerationJobSummary_template - The Amazon Resource Name (arn) of the world template.

worldGenerationJobSummary_status :: Lens' WorldGenerationJobSummary (Maybe WorldGenerationJobStatus)

The status of the world generator job:

Pending

The world generator job request is pending.

Running

The world generator job is running.

Completed

The world generator job completed.

Failed

The world generator job failed. See failureCode for more information.

PartialFailed

Some worlds did not generate.

Canceled

The world generator job was cancelled.

Canceling

The world generator job is being cancelled.

worldGenerationJobSummary_arn :: Lens' WorldGenerationJobSummary (Maybe Text)

The Amazon Resource Name (ARN) of the world generator job.

worldGenerationJobSummary_createdAt :: Lens' WorldGenerationJobSummary (Maybe UTCTime)

The time, in milliseconds since the epoch, when the world generator job was created.

worldGenerationJobSummary_worldCount :: Lens' WorldGenerationJobSummary (Maybe WorldCount)

Information about the world count.

worldGenerationJobSummary_succeededWorldCount :: Lens' WorldGenerationJobSummary (Maybe Int)

The number of worlds that were generated.

worldGenerationJobSummary_failedWorldCount :: Lens' WorldGenerationJobSummary (Maybe Int)

The number of worlds that failed.

worldGenerationJobSummary_template :: Lens' WorldGenerationJobSummary (Maybe Text)

The Amazon Resource Name (arn) of the world template.