Amazonka.RobOMaker.Types.WorldSummary

Description

Documentation

data WorldSummary

Information about a world.

See: newWorldSummary smart constructor.

Constructors

WorldSummary'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the world.

createdAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the world was created.

template :: Maybe Text

The Amazon Resource Name (arn) of the world template.

generationJob :: Maybe Text

The Amazon Resource Name (arn) of the world generation job.

newWorldSummary :: WorldSummary

Create a value of WorldSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:WorldSummary', worldSummary_arn - The Amazon Resource Name (ARN) of the world.

$sel:createdAt:WorldSummary', worldSummary_createdAt - The time, in milliseconds since the epoch, when the world was created.

$sel:template:WorldSummary', worldSummary_template - The Amazon Resource Name (arn) of the world template.

$sel:generationJob:WorldSummary', worldSummary_generationJob - The Amazon Resource Name (arn) of the world generation job.

worldSummary_arn :: Lens' WorldSummary (Maybe Text)

The Amazon Resource Name (ARN) of the world.

worldSummary_createdAt :: Lens' WorldSummary (Maybe UTCTime)

The time, in milliseconds since the epoch, when the world was created.

worldSummary_template :: Lens' WorldSummary (Maybe Text)

The Amazon Resource Name (arn) of the world template.

worldSummary_generationJob :: Lens' WorldSummary (Maybe Text)

The Amazon Resource Name (arn) of the world generation job.