Amazonka.RobOMaker.UpdateRobotApplication

Description

Updates a robot application.

Creating a Request

data UpdateRobotApplication

See: newUpdateRobotApplication smart constructor.

Constructors

UpdateRobotApplication'

Fields

currentRevisionId :: Maybe Text

The revision id for the robot application.

environment :: Maybe Environment

The object that contains the Docker image URI for your robot application.

sources :: Maybe [SourceConfig]

The sources of the robot application.

application :: Text

The application information for the robot application.

robotSoftwareSuite :: RobotSoftwareSuite

The robot software suite (ROS distribution) used by the robot application.

newUpdateRobotApplication

Arguments

:: Text
$sel:application:UpdateRobotApplication'

-> RobotSoftwareSuite
$sel:robotSoftwareSuite:UpdateRobotApplication'

-> UpdateRobotApplication

Create a value of UpdateRobotApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentRevisionId:UpdateRobotApplication', updateRobotApplication_currentRevisionId - The revision id for the robot application.

$sel:environment:UpdateRobotApplication', updateRobotApplication_environment - The object that contains the Docker image URI for your robot application.

$sel:sources:UpdateRobotApplication', updateRobotApplication_sources - The sources of the robot application.

$sel:application:UpdateRobotApplication', updateRobotApplication_application - The application information for the robot application.

$sel:robotSoftwareSuite:UpdateRobotApplication', updateRobotApplication_robotSoftwareSuite - The robot software suite (ROS distribution) used by the robot application.

Request Lenses

updateRobotApplication_currentRevisionId :: Lens' UpdateRobotApplication (Maybe Text)

The revision id for the robot application.

updateRobotApplication_environment :: Lens' UpdateRobotApplication (Maybe Environment)

The object that contains the Docker image URI for your robot application.

updateRobotApplication_sources :: Lens' UpdateRobotApplication (Maybe [SourceConfig])

The sources of the robot application.

updateRobotApplication_application :: Lens' UpdateRobotApplication Text

The application information for the robot application.

updateRobotApplication_robotSoftwareSuite :: Lens' UpdateRobotApplication RobotSoftwareSuite

The robot software suite (ROS distribution) used by the robot application.

Destructuring the Response

data UpdateRobotApplicationResponse

See: newUpdateRobotApplicationResponse smart constructor.

Constructors

UpdateRobotApplicationResponse'

Fields

lastUpdatedAt :: Maybe POSIX

The time, in milliseconds since the epoch, when the robot application was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the updated robot application.

environment :: Maybe Environment

The object that contains the Docker image URI for your robot application.

sources :: Maybe [Source]

The sources of the robot application.

name :: Maybe Text

The name of the robot application.

version :: Maybe Text

The version of the robot application.

robotSoftwareSuite :: Maybe RobotSoftwareSuite

The robot software suite (ROS distribution) used by the robot application.

revisionId :: Maybe Text

The revision id of the robot application.

httpStatus :: Int

The response's http status code.

newUpdateRobotApplicationResponse

Arguments

:: Int
$sel:httpStatus:UpdateRobotApplicationResponse'

-> UpdateRobotApplicationResponse

Create a value of UpdateRobotApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedAt:UpdateRobotApplicationResponse', updateRobotApplicationResponse_lastUpdatedAt - The time, in milliseconds since the epoch, when the robot application was last updated.

$sel:arn:UpdateRobotApplicationResponse', updateRobotApplicationResponse_arn - The Amazon Resource Name (ARN) of the updated robot application.

$sel:environment:UpdateRobotApplication', updateRobotApplicationResponse_environment - The object that contains the Docker image URI for your robot application.

$sel:sources:UpdateRobotApplication', updateRobotApplicationResponse_sources - The sources of the robot application.

$sel:name:UpdateRobotApplicationResponse', updateRobotApplicationResponse_name - The name of the robot application.

$sel:version:UpdateRobotApplicationResponse', updateRobotApplicationResponse_version - The version of the robot application.

$sel:robotSoftwareSuite:UpdateRobotApplication', updateRobotApplicationResponse_robotSoftwareSuite - The robot software suite (ROS distribution) used by the robot application.

$sel:revisionId:UpdateRobotApplicationResponse', updateRobotApplicationResponse_revisionId - The revision id of the robot application.

$sel:httpStatus:UpdateRobotApplicationResponse', updateRobotApplicationResponse_httpStatus - The response's http status code.

Response Lenses

updateRobotApplicationResponse_lastUpdatedAt :: Lens' UpdateRobotApplicationResponse (Maybe UTCTime)

The time, in milliseconds since the epoch, when the robot application was last updated.

updateRobotApplicationResponse_arn :: Lens' UpdateRobotApplicationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the updated robot application.

updateRobotApplicationResponse_environment :: Lens' UpdateRobotApplicationResponse (Maybe Environment)

The object that contains the Docker image URI for your robot application.

updateRobotApplicationResponse_sources :: Lens' UpdateRobotApplicationResponse (Maybe [Source])

The sources of the robot application.

updateRobotApplicationResponse_name :: Lens' UpdateRobotApplicationResponse (Maybe Text)

The name of the robot application.

updateRobotApplicationResponse_version :: Lens' UpdateRobotApplicationResponse (Maybe Text)

The version of the robot application.

updateRobotApplicationResponse_robotSoftwareSuite :: Lens' UpdateRobotApplicationResponse (Maybe RobotSoftwareSuite)

The robot software suite (ROS distribution) used by the robot application.

updateRobotApplicationResponse_revisionId :: Lens' UpdateRobotApplicationResponse (Maybe Text)

The revision id of the robot application.

updateRobotApplicationResponse_httpStatus :: Lens' UpdateRobotApplicationResponse Int

The response's http status code.