Amazonka.Route53.CreateHealthCheck

Description

Creates a new health check.

For information about adding health checks to resource record sets, see HealthCheckId in ChangeResourceRecordSets.

ELB Load Balancers

If you're registering EC2 instances with an Elastic Load Balancing (ELB) load balancer, do not create Amazon Route 53 health checks for the EC2 instances. When you register an EC2 instance with a load balancer, you configure settings for an ELB health check, which performs a similar function to a Route 53 health check.

Private Hosted Zones

You can associate health checks with failover resource record sets in a private hosted zone. Note the following:

Route 53 health checkers are outside the VPC. To check the health of an endpoint within a VPC by IP address, you must assign a public IP address to the instance in the VPC.

You can configure a health checker to check the health of an external resource that the instance relies on, such as a database server.

You can create a CloudWatch metric, associate an alarm with the metric, and then create a health check that is based on the state of the alarm. For example, you might create a CloudWatch metric that checks the status of the Amazon EC2 StatusCheckFailed metric, add an alarm to the metric, and then create a health check that is based on the state of the alarm. For information about creating CloudWatch metrics and alarms by using the CloudWatch console, see the Amazon CloudWatch User Guide.

Creating a Request

data CreateHealthCheck

A complex type that contains the health check request information.

See: newCreateHealthCheck smart constructor.

Constructors

CreateHealthCheck'

Fields

callerReference :: Text

A unique string that identifies the request and that allows you to retry a failed CreateHealthCheck request without the risk of creating two identical health checks:

If you send a CreateHealthCheck request with the same CallerReference and settings as a previous request, and if the health check doesn't exist, Amazon Route 53 creates the health check. If the health check does exist, Route 53 returns the settings for the existing health check.

If you send a CreateHealthCheck request with the same CallerReference as a deleted health check, regardless of the settings, Route 53 returns a HealthCheckAlreadyExists error.

If you send a CreateHealthCheck request with the same CallerReference as an existing health check but with different settings, Route 53 returns a HealthCheckAlreadyExists error.

If you send a CreateHealthCheck request with a unique CallerReference but settings identical to an existing health check, Route 53 creates the health check.

healthCheckConfig :: HealthCheckConfig

A complex type that contains settings for a new health check.

newCreateHealthCheck

Arguments

:: Text
$sel:callerReference:CreateHealthCheck'

-> HealthCheckConfig
$sel:healthCheckConfig:CreateHealthCheck'

-> CreateHealthCheck

Create a value of CreateHealthCheck with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:callerReference:CreateHealthCheck', createHealthCheck_callerReference - A unique string that identifies the request and that allows you to retry a failed CreateHealthCheck request without the risk of creating two identical health checks:

If you send a CreateHealthCheck request with the same CallerReference and settings as a previous request, and if the health check doesn't exist, Amazon Route 53 creates the health check. If the health check does exist, Route 53 returns the settings for the existing health check.

If you send a CreateHealthCheck request with the same CallerReference as a deleted health check, regardless of the settings, Route 53 returns a HealthCheckAlreadyExists error.

If you send a CreateHealthCheck request with the same CallerReference as an existing health check but with different settings, Route 53 returns a HealthCheckAlreadyExists error.

If you send a CreateHealthCheck request with a unique CallerReference but settings identical to an existing health check, Route 53 creates the health check.

$sel:healthCheckConfig:CreateHealthCheck', createHealthCheck_healthCheckConfig - A complex type that contains settings for a new health check.

Request Lenses

createHealthCheck_callerReference :: Lens' CreateHealthCheck Text

A unique string that identifies the request and that allows you to retry a failed CreateHealthCheck request without the risk of creating two identical health checks:

If you send a CreateHealthCheck request with the same CallerReference and settings as a previous request, and if the health check doesn't exist, Amazon Route 53 creates the health check. If the health check does exist, Route 53 returns the settings for the existing health check.

If you send a CreateHealthCheck request with the same CallerReference as a deleted health check, regardless of the settings, Route 53 returns a HealthCheckAlreadyExists error.

If you send a CreateHealthCheck request with the same CallerReference as an existing health check but with different settings, Route 53 returns a HealthCheckAlreadyExists error.

If you send a CreateHealthCheck request with a unique CallerReference but settings identical to an existing health check, Route 53 creates the health check.

createHealthCheck_healthCheckConfig :: Lens' CreateHealthCheck HealthCheckConfig

A complex type that contains settings for a new health check.

Destructuring the Response

data CreateHealthCheckResponse

A complex type containing the response information for the new health check.

See: newCreateHealthCheckResponse smart constructor.

Constructors

CreateHealthCheckResponse'

Fields

httpStatus :: Int

The response's http status code.

healthCheck :: HealthCheck

A complex type that contains identifying information about the health check.

location :: Text

The unique URL representing the new health check.

newCreateHealthCheckResponse

Arguments

:: Int
$sel:httpStatus:CreateHealthCheckResponse'

-> HealthCheck
$sel:healthCheck:CreateHealthCheckResponse'

-> Text
$sel:location:CreateHealthCheckResponse'

-> CreateHealthCheckResponse

Create a value of CreateHealthCheckResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateHealthCheckResponse', createHealthCheckResponse_httpStatus - The response's http status code.

$sel:healthCheck:CreateHealthCheckResponse', createHealthCheckResponse_healthCheck - A complex type that contains identifying information about the health check.

$sel:location:CreateHealthCheckResponse', createHealthCheckResponse_location - The unique URL representing the new health check.

Response Lenses

createHealthCheckResponse_httpStatus :: Lens' CreateHealthCheckResponse Int

The response's http status code.

createHealthCheckResponse_healthCheck :: Lens' CreateHealthCheckResponse HealthCheck

A complex type that contains identifying information about the health check.

createHealthCheckResponse_location :: Lens' CreateHealthCheckResponse Text

The unique URL representing the new health check.