Amazonka.Route53.CreateReusableDelegationSet

Description

Creates a delegation set (a group of four name servers) that can be reused by multiple hosted zones that were created by the same Amazon Web Services account.

You can also create a reusable delegation set that uses the four name servers that are associated with an existing hosted zone. Specify the hosted zone ID in the CreateReusableDelegationSet request.

You can't associate a reusable delegation set with a private hosted zone.

For information about using a reusable delegation set to configure white label name servers, see Configuring White Label Name Servers.

The process for migrating existing hosted zones to use a reusable delegation set is comparable to the process for configuring white label name servers. You need to perform the following steps:

Create a reusable delegation set.

Recreate hosted zones, and reduce the TTL to 60 seconds or less.

Recreate resource record sets in the new hosted zones.

Change the registrar's name servers to use the name servers for the new hosted zones.

Monitor traffic for the website or application.

Change TTLs back to their original values.

If you want to migrate existing hosted zones to use a reusable delegation set, the existing hosted zones can't use any of the name servers that are assigned to the reusable delegation set. If one or more hosted zones do use one or more name servers that are assigned to the reusable delegation set, you can do one of the following:

For small numbers of hosted zones—up to a few hundred—it's relatively easy to create reusable delegation sets until you get one that has four name servers that don't overlap with any of the name servers in your hosted zones.

For larger numbers of hosted zones, the easiest solution is to use more than one reusable delegation set.

For larger numbers of hosted zones, you can also migrate hosted zones that have overlapping name servers to hosted zones that don't have overlapping name servers, then migrate the hosted zones again to use the reusable delegation set.

Creating a Request

data CreateReusableDelegationSet

See: newCreateReusableDelegationSet smart constructor.

Constructors

CreateReusableDelegationSet'

Fields

hostedZoneId :: Maybe ResourceId

If you want to mark the delegation set for an existing hosted zone as reusable, the ID for that hosted zone.

callerReference :: Text

A unique string that identifies the request, and that allows you to retry failed CreateReusableDelegationSet requests without the risk of executing the operation twice. You must use a unique CallerReference string every time you submit a CreateReusableDelegationSet request. CallerReference can be any unique string, for example a date/time stamp.

newCreateReusableDelegationSet

Arguments

:: Text
$sel:callerReference:CreateReusableDelegationSet'

-> CreateReusableDelegationSet

Create a value of CreateReusableDelegationSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostedZoneId:CreateReusableDelegationSet', createReusableDelegationSet_hostedZoneId - If you want to mark the delegation set for an existing hosted zone as reusable, the ID for that hosted zone.

$sel:callerReference:CreateReusableDelegationSet', createReusableDelegationSet_callerReference - A unique string that identifies the request, and that allows you to retry failed CreateReusableDelegationSet requests without the risk of executing the operation twice. You must use a unique CallerReference string every time you submit a CreateReusableDelegationSet request. CallerReference can be any unique string, for example a date/time stamp.

Request Lenses

createReusableDelegationSet_hostedZoneId :: Lens' CreateReusableDelegationSet (Maybe ResourceId)

If you want to mark the delegation set for an existing hosted zone as reusable, the ID for that hosted zone.

createReusableDelegationSet_callerReference :: Lens' CreateReusableDelegationSet Text

A unique string that identifies the request, and that allows you to retry failed CreateReusableDelegationSet requests without the risk of executing the operation twice. You must use a unique CallerReference string every time you submit a CreateReusableDelegationSet request. CallerReference can be any unique string, for example a date/time stamp.

Destructuring the Response

data CreateReusableDelegationSetResponse

See: newCreateReusableDelegationSetResponse smart constructor.

Constructors

CreateReusableDelegationSetResponse'

Fields

httpStatus :: Int

The response's http status code.

delegationSet :: DelegationSet

A complex type that contains name server information.

location :: Text

The unique URL representing the new reusable delegation set.

newCreateReusableDelegationSetResponse

Arguments

:: Int
$sel:httpStatus:CreateReusableDelegationSetResponse'

-> DelegationSet
$sel:delegationSet:CreateReusableDelegationSetResponse'

-> Text
$sel:location:CreateReusableDelegationSetResponse'

-> CreateReusableDelegationSetResponse

Create a value of CreateReusableDelegationSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateReusableDelegationSetResponse', createReusableDelegationSetResponse_httpStatus - The response's http status code.

$sel:delegationSet:CreateReusableDelegationSetResponse', createReusableDelegationSetResponse_delegationSet - A complex type that contains name server information.

$sel:location:CreateReusableDelegationSetResponse', createReusableDelegationSetResponse_location - The unique URL representing the new reusable delegation set.

Response Lenses

createReusableDelegationSetResponse_httpStatus :: Lens' CreateReusableDelegationSetResponse Int

The response's http status code.

createReusableDelegationSetResponse_delegationSet :: Lens' CreateReusableDelegationSetResponse DelegationSet

A complex type that contains name server information.

createReusableDelegationSetResponse_location :: Lens' CreateReusableDelegationSetResponse Text

The unique URL representing the new reusable delegation set.