Amazonka.Route53.CreateTrafficPolicy

Description

Creates a traffic policy, which you use to create multiple DNS resource record sets for one domain name (such as example.com) or one subdomain name (such as www.example.com).

Creating a Request

data CreateTrafficPolicy

A complex type that contains information about the traffic policy that you want to create.

See: newCreateTrafficPolicy smart constructor.

Constructors

CreateTrafficPolicy'

Fields

comment :: Maybe Text

(Optional) Any comments that you want to include about the traffic policy.

name :: Text

The name of the traffic policy.

document :: Text

The definition of this traffic policy in JSON format. For more information, see Traffic Policy Document Format.

newCreateTrafficPolicy

Arguments

:: Text
$sel:name:CreateTrafficPolicy'

-> Text
$sel:document:CreateTrafficPolicy'

-> CreateTrafficPolicy

Create a value of CreateTrafficPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comment:CreateTrafficPolicy', createTrafficPolicy_comment - (Optional) Any comments that you want to include about the traffic policy.

$sel:name:CreateTrafficPolicy', createTrafficPolicy_name - The name of the traffic policy.

$sel:document:CreateTrafficPolicy', createTrafficPolicy_document - The definition of this traffic policy in JSON format. For more information, see Traffic Policy Document Format.

Request Lenses

createTrafficPolicy_comment :: Lens' CreateTrafficPolicy (Maybe Text)

(Optional) Any comments that you want to include about the traffic policy.

createTrafficPolicy_name :: Lens' CreateTrafficPolicy Text

The name of the traffic policy.

createTrafficPolicy_document :: Lens' CreateTrafficPolicy Text

The definition of this traffic policy in JSON format. For more information, see Traffic Policy Document Format.

Destructuring the Response

data CreateTrafficPolicyResponse

A complex type that contains the response information for the CreateTrafficPolicy request.

See: newCreateTrafficPolicyResponse smart constructor.

Constructors

CreateTrafficPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

trafficPolicy :: TrafficPolicy

A complex type that contains settings for the new traffic policy.

location :: Text

A unique URL that represents a new traffic policy.

newCreateTrafficPolicyResponse

Arguments

:: Int
$sel:httpStatus:CreateTrafficPolicyResponse'

-> TrafficPolicy
$sel:trafficPolicy:CreateTrafficPolicyResponse'

-> Text
$sel:location:CreateTrafficPolicyResponse'

-> CreateTrafficPolicyResponse

Create a value of CreateTrafficPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateTrafficPolicyResponse', createTrafficPolicyResponse_httpStatus - The response's http status code.

$sel:trafficPolicy:CreateTrafficPolicyResponse', createTrafficPolicyResponse_trafficPolicy - A complex type that contains settings for the new traffic policy.

$sel:location:CreateTrafficPolicyResponse', createTrafficPolicyResponse_location - A unique URL that represents a new traffic policy.

Response Lenses

createTrafficPolicyResponse_httpStatus :: Lens' CreateTrafficPolicyResponse Int

The response's http status code.

createTrafficPolicyResponse_trafficPolicy :: Lens' CreateTrafficPolicyResponse TrafficPolicy

A complex type that contains settings for the new traffic policy.

createTrafficPolicyResponse_location :: Lens' CreateTrafficPolicyResponse Text

A unique URL that represents a new traffic policy.