Amazonka.Route53.DeleteQueryLoggingConfig

Description

Deletes a configuration for DNS query logging. If you delete a configuration, Amazon Route 53 stops sending query logs to CloudWatch Logs. Route 53 doesn't delete any logs that are already in CloudWatch Logs.

For more information about DNS query logs, see CreateQueryLoggingConfig.

Creating a Request

data DeleteQueryLoggingConfig

See: newDeleteQueryLoggingConfig smart constructor.

Constructors

DeleteQueryLoggingConfig'

Fields

id :: Text

The ID of the configuration that you want to delete.

newDeleteQueryLoggingConfig

Arguments

:: Text
$sel:id:DeleteQueryLoggingConfig'

-> DeleteQueryLoggingConfig

Create a value of DeleteQueryLoggingConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeleteQueryLoggingConfig', deleteQueryLoggingConfig_id - The ID of the configuration that you want to delete.

Request Lenses

deleteQueryLoggingConfig_id :: Lens' DeleteQueryLoggingConfig Text

The ID of the configuration that you want to delete.

Destructuring the Response

data DeleteQueryLoggingConfigResponse

See: newDeleteQueryLoggingConfigResponse smart constructor.

Constructors

DeleteQueryLoggingConfigResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteQueryLoggingConfigResponse

Arguments

:: Int
$sel:httpStatus:DeleteQueryLoggingConfigResponse'

-> DeleteQueryLoggingConfigResponse

Create a value of DeleteQueryLoggingConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteQueryLoggingConfigResponse', deleteQueryLoggingConfigResponse_httpStatus - The response's http status code.

Response Lenses

deleteQueryLoggingConfigResponse_httpStatus :: Lens' DeleteQueryLoggingConfigResponse Int

The response's http status code.