Amazonka.Route53.DeleteReusableDelegationSet

Description

Deletes a reusable delegation set.

You can delete a reusable delegation set only if it isn't associated with any hosted zones.

To verify that the reusable delegation set is not associated with any hosted zones, submit a GetReusableDelegationSet request and specify the ID of the reusable delegation set that you want to delete.

Creating a Request

data DeleteReusableDelegationSet

A request to delete a reusable delegation set.

See: newDeleteReusableDelegationSet smart constructor.

Constructors

DeleteReusableDelegationSet'

Fields

id :: ResourceId

The ID of the reusable delegation set that you want to delete.

newDeleteReusableDelegationSet

Arguments

:: ResourceId
$sel:id:DeleteReusableDelegationSet'

-> DeleteReusableDelegationSet

Create a value of DeleteReusableDelegationSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DeleteReusableDelegationSet', deleteReusableDelegationSet_id - The ID of the reusable delegation set that you want to delete.

Request Lenses

deleteReusableDelegationSet_id :: Lens' DeleteReusableDelegationSet ResourceId

The ID of the reusable delegation set that you want to delete.

Destructuring the Response

data DeleteReusableDelegationSetResponse

An empty element.

See: newDeleteReusableDelegationSetResponse smart constructor.

Constructors

DeleteReusableDelegationSetResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteReusableDelegationSetResponse

Arguments

:: Int
$sel:httpStatus:DeleteReusableDelegationSetResponse'

-> DeleteReusableDelegationSetResponse

Create a value of DeleteReusableDelegationSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteReusableDelegationSetResponse', deleteReusableDelegationSetResponse_httpStatus - The response's http status code.

Response Lenses

deleteReusableDelegationSetResponse_httpStatus :: Lens' DeleteReusableDelegationSetResponse Int

The response's http status code.