Amazonka.Route53.GetGeoLocation

Description

Gets information about whether a specified geographic location is supported for Amazon Route 53 geolocation resource record sets.

Route 53 does not perform authorization for this API because it retrieves information that is already available to the public.

Use the following syntax to determine whether a continent is supported for geolocation:

GET /2013-04-01/geolocation?continentcode=two-letter abbreviation for a continent

Use the following syntax to determine whether a country is supported for geolocation:

GET /2013-04-01/geolocation?countrycode=two-character country code

Use the following syntax to determine whether a subdivision of a country is supported for geolocation:

GET /2013-04-01/geolocation?countrycode=two-character country code&subdivisioncode=subdivision code

Creating a Request

data GetGeoLocation

A request for information about whether a specified geographic location is supported for Amazon Route 53 geolocation resource record sets.

See: newGetGeoLocation smart constructor.

Constructors

GetGeoLocation'

Fields

subdivisionCode :: Maybe Text

The code for the subdivision, such as a particular state within the United States. For a list of US state abbreviations, see Appendix B: Two–Letter State and Possession Abbreviations on the United States Postal Service website. For a list of all supported subdivision codes, use the ListGeoLocations API.

countryCode :: Maybe Text

Amazon Route 53 uses the two-letter country codes that are specified in ISO standard 3166-1 alpha-2.

continentCode :: Maybe Text

For geolocation resource record sets, a two-letter abbreviation that identifies a continent. Amazon Route 53 supports the following continent codes:

AF: Africa

AN: Antarctica

AS: Asia

EU: Europe

OC: Oceania

NA: North America

SA: South America

newGetGeoLocation :: GetGeoLocation

Create a value of GetGeoLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subdivisionCode:GetGeoLocation', getGeoLocation_subdivisionCode - The code for the subdivision, such as a particular state within the United States. For a list of US state abbreviations, see Appendix B: Two–Letter State and Possession Abbreviations on the United States Postal Service website. For a list of all supported subdivision codes, use the ListGeoLocations API.

$sel:countryCode:GetGeoLocation', getGeoLocation_countryCode - Amazon Route 53 uses the two-letter country codes that are specified in ISO standard 3166-1 alpha-2.

$sel:continentCode:GetGeoLocation', getGeoLocation_continentCode - For geolocation resource record sets, a two-letter abbreviation that identifies a continent. Amazon Route 53 supports the following continent codes:

AF: Africa

AN: Antarctica

AS: Asia

EU: Europe

OC: Oceania

NA: North America

SA: South America

Request Lenses

getGeoLocation_subdivisionCode :: Lens' GetGeoLocation (Maybe Text)

The code for the subdivision, such as a particular state within the United States. For a list of US state abbreviations, see Appendix B: Two–Letter State and Possession Abbreviations on the United States Postal Service website. For a list of all supported subdivision codes, use the ListGeoLocations API.

getGeoLocation_countryCode :: Lens' GetGeoLocation (Maybe Text)

Amazon Route 53 uses the two-letter country codes that are specified in ISO standard 3166-1 alpha-2.

getGeoLocation_continentCode :: Lens' GetGeoLocation (Maybe Text)

For geolocation resource record sets, a two-letter abbreviation that identifies a continent. Amazon Route 53 supports the following continent codes:

AF: Africa

AN: Antarctica

AS: Asia

EU: Europe

OC: Oceania

NA: North America

SA: South America

Destructuring the Response

data GetGeoLocationResponse

A complex type that contains the response information for the specified geolocation code.

See: newGetGeoLocationResponse smart constructor.

Constructors

GetGeoLocationResponse'

Fields

httpStatus :: Int

The response's http status code.

geoLocationDetails :: GeoLocationDetails

A complex type that contains the codes and full continent, country, and subdivision names for the specified geolocation code.

newGetGeoLocationResponse

Arguments

:: Int
$sel:httpStatus:GetGeoLocationResponse'

-> GeoLocationDetails
$sel:geoLocationDetails:GetGeoLocationResponse'

-> GetGeoLocationResponse

Create a value of GetGeoLocationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetGeoLocationResponse', getGeoLocationResponse_httpStatus - The response's http status code.

$sel:geoLocationDetails:GetGeoLocationResponse', getGeoLocationResponse_geoLocationDetails - A complex type that contains the codes and full continent, country, and subdivision names for the specified geolocation code.

Response Lenses

getGeoLocationResponse_httpStatus :: Lens' GetGeoLocationResponse Int

The response's http status code.

getGeoLocationResponse_geoLocationDetails :: Lens' GetGeoLocationResponse GeoLocationDetails

A complex type that contains the codes and full continent, country, and subdivision names for the specified geolocation code.