Amazonka.Route53.GetHealthCheckStatus

Description

Gets status of a specified health check.

This API is intended for use during development to diagnose behavior. It doesn’t support production use-cases with high query rates that require immediate and actionable responses.

Creating a Request

data GetHealthCheckStatus

A request to get the status for a health check.

See: newGetHealthCheckStatus smart constructor.

Constructors

GetHealthCheckStatus'

Fields

healthCheckId :: Text

The ID for the health check that you want the current status for. When you created the health check, CreateHealthCheck returned the ID in the response, in the HealthCheckId element.

If you want to check the status of a calculated health check, you must use the Amazon Route 53 console or the CloudWatch console. You can't use GetHealthCheckStatus to get the status of a calculated health check.

newGetHealthCheckStatus

Arguments

:: Text
$sel:healthCheckId:GetHealthCheckStatus'

-> GetHealthCheckStatus

Create a value of GetHealthCheckStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:healthCheckId:GetHealthCheckStatus', getHealthCheckStatus_healthCheckId - The ID for the health check that you want the current status for. When you created the health check, CreateHealthCheck returned the ID in the response, in the HealthCheckId element.

If you want to check the status of a calculated health check, you must use the Amazon Route 53 console or the CloudWatch console. You can't use GetHealthCheckStatus to get the status of a calculated health check.

Request Lenses

getHealthCheckStatus_healthCheckId :: Lens' GetHealthCheckStatus Text

The ID for the health check that you want the current status for. When you created the health check, CreateHealthCheck returned the ID in the response, in the HealthCheckId element.

If you want to check the status of a calculated health check, you must use the Amazon Route 53 console or the CloudWatch console. You can't use GetHealthCheckStatus to get the status of a calculated health check.

Destructuring the Response

data GetHealthCheckStatusResponse

A complex type that contains the response to a GetHealthCheck request.

See: newGetHealthCheckStatusResponse smart constructor.

Constructors

GetHealthCheckStatusResponse'

Fields

httpStatus :: Int

The response's http status code.

healthCheckObservations :: [HealthCheckObservation]

A list that contains one HealthCheckObservation element for each Amazon Route 53 health checker that is reporting a status about the health check endpoint.

newGetHealthCheckStatusResponse

Arguments

:: Int
$sel:httpStatus:GetHealthCheckStatusResponse'

-> GetHealthCheckStatusResponse

Create a value of GetHealthCheckStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetHealthCheckStatusResponse', getHealthCheckStatusResponse_httpStatus - The response's http status code.

$sel:healthCheckObservations:GetHealthCheckStatusResponse', getHealthCheckStatusResponse_healthCheckObservations - A list that contains one HealthCheckObservation element for each Amazon Route 53 health checker that is reporting a status about the health check endpoint.

Response Lenses

getHealthCheckStatusResponse_httpStatus :: Lens' GetHealthCheckStatusResponse Int

The response's http status code.

getHealthCheckStatusResponse_healthCheckObservations :: Lens' GetHealthCheckStatusResponse [HealthCheckObservation]

A list that contains one HealthCheckObservation element for each Amazon Route 53 health checker that is reporting a status about the health check endpoint.