Amazonka.Route53.GetHostedZoneLimit

Description

Gets the specified limit for a specified hosted zone, for example, the maximum number of records that you can create in the hosted zone.

For the default limit, see Limits in the Amazon Route 53 Developer Guide. To request a higher limit, open a case.

Creating a Request

data GetHostedZoneLimit

A complex type that contains information about the request to create a hosted zone.

See: newGetHostedZoneLimit smart constructor.

Constructors

GetHostedZoneLimit'

Fields

type' :: HostedZoneLimitType

The limit that you want to get. Valid values include the following:

MAX_RRSETS_BY_ZONE: The maximum number of records that you can create in the specified hosted zone.

MAX_VPCS_ASSOCIATED_BY_ZONE: The maximum number of Amazon VPCs that you can associate with the specified private hosted zone.

hostedZoneId :: ResourceId

The ID of the hosted zone that you want to get a limit for.

newGetHostedZoneLimit

Arguments

:: HostedZoneLimitType
$sel:type':GetHostedZoneLimit'

-> ResourceId
$sel:hostedZoneId:GetHostedZoneLimit'

-> GetHostedZoneLimit

Create a value of GetHostedZoneLimit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':GetHostedZoneLimit', getHostedZoneLimit_type - The limit that you want to get. Valid values include the following:

MAX_RRSETS_BY_ZONE: The maximum number of records that you can create in the specified hosted zone.

MAX_VPCS_ASSOCIATED_BY_ZONE: The maximum number of Amazon VPCs that you can associate with the specified private hosted zone.

$sel:hostedZoneId:GetHostedZoneLimit', getHostedZoneLimit_hostedZoneId - The ID of the hosted zone that you want to get a limit for.

Request Lenses

getHostedZoneLimit_type :: Lens' GetHostedZoneLimit HostedZoneLimitType

The limit that you want to get. Valid values include the following:

MAX_RRSETS_BY_ZONE: The maximum number of records that you can create in the specified hosted zone.

MAX_VPCS_ASSOCIATED_BY_ZONE: The maximum number of Amazon VPCs that you can associate with the specified private hosted zone.

getHostedZoneLimit_hostedZoneId :: Lens' GetHostedZoneLimit ResourceId

The ID of the hosted zone that you want to get a limit for.

Destructuring the Response

data GetHostedZoneLimitResponse

A complex type that contains the requested limit.

See: newGetHostedZoneLimitResponse smart constructor.

Constructors

GetHostedZoneLimitResponse'

Fields

httpStatus :: Int

The response's http status code.

limit :: HostedZoneLimit

The current setting for the specified limit. For example, if you specified MAX_RRSETS_BY_ZONE for the value of Type in the request, the value of Limit is the maximum number of records that you can create in the specified hosted zone.

count :: Natural

The current number of entities that you have created of the specified type. For example, if you specified MAX_RRSETS_BY_ZONE for the value of Type in the request, the value of Count is the current number of records that you have created in the specified hosted zone.

newGetHostedZoneLimitResponse

Arguments

:: Int
$sel:httpStatus:GetHostedZoneLimitResponse'

-> HostedZoneLimit
$sel:limit:GetHostedZoneLimitResponse'

-> Natural
$sel:count:GetHostedZoneLimitResponse'

-> GetHostedZoneLimitResponse

Create a value of GetHostedZoneLimitResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetHostedZoneLimitResponse', getHostedZoneLimitResponse_httpStatus - The response's http status code.

$sel:limit:GetHostedZoneLimitResponse', getHostedZoneLimitResponse_limit - The current setting for the specified limit. For example, if you specified MAX_RRSETS_BY_ZONE for the value of Type in the request, the value of Limit is the maximum number of records that you can create in the specified hosted zone.

$sel:count:GetHostedZoneLimitResponse', getHostedZoneLimitResponse_count - The current number of entities that you have created of the specified type. For example, if you specified MAX_RRSETS_BY_ZONE for the value of Type in the request, the value of Count is the current number of records that you have created in the specified hosted zone.

Response Lenses

getHostedZoneLimitResponse_httpStatus :: Lens' GetHostedZoneLimitResponse Int

The response's http status code.

getHostedZoneLimitResponse_limit :: Lens' GetHostedZoneLimitResponse HostedZoneLimit

The current setting for the specified limit. For example, if you specified MAX_RRSETS_BY_ZONE for the value of Type in the request, the value of Limit is the maximum number of records that you can create in the specified hosted zone.

getHostedZoneLimitResponse_count :: Lens' GetHostedZoneLimitResponse Natural

The current number of entities that you have created of the specified type. For example, if you specified MAX_RRSETS_BY_ZONE for the value of Type in the request, the value of Count is the current number of records that you have created in the specified hosted zone.