Amazonka.Route53.ListTrafficPolicyInstances

Description

Gets information about the traffic policy instances that you created by using the current Amazon Web Services account.

After you submit an UpdateTrafficPolicyInstance request, there's a brief delay while Amazon Route 53 creates the resource record sets that are specified in the traffic policy definition. For more information, see the State response element.

Route 53 returns a maximum of 100 items in each response. If you have a lot of traffic policy instances, you can use the MaxItems parameter to list them in groups of up to 100.

Creating a Request

data ListTrafficPolicyInstances

A request to get information about the traffic policy instances that you created by using the current Amazon Web Services account.

See: newListTrafficPolicyInstances smart constructor.

Constructors

ListTrafficPolicyInstances'

Fields

trafficPolicyInstanceTypeMarker :: Maybe RRType

If the value of IsTruncated in the previous response was true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancetype, specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the type of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

maxItems :: Maybe Text

The maximum number of traffic policy instances that you want Amazon Route 53 to return in response to a ListTrafficPolicyInstances request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true, and the values of HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance in the next group of MaxItems traffic policy instances.

hostedZoneIdMarker :: Maybe ResourceId

If the value of IsTruncated in the previous response was true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of HostedZoneId, specify the value of HostedZoneIdMarker from the previous response, which is the hosted zone ID of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

trafficPolicyInstanceNameMarker :: Maybe Text

If the value of IsTruncated in the previous response was true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancename, specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

newListTrafficPolicyInstances :: ListTrafficPolicyInstances

Create a value of ListTrafficPolicyInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficPolicyInstanceTypeMarker:ListTrafficPolicyInstances', listTrafficPolicyInstances_trafficPolicyInstanceTypeMarker - If the value of IsTruncated in the previous response was true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancetype, specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the type of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

$sel:maxItems:ListTrafficPolicyInstances', listTrafficPolicyInstances_maxItems - The maximum number of traffic policy instances that you want Amazon Route 53 to return in response to a ListTrafficPolicyInstances request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true, and the values of HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance in the next group of MaxItems traffic policy instances.

$sel:hostedZoneIdMarker:ListTrafficPolicyInstances', listTrafficPolicyInstances_hostedZoneIdMarker - If the value of IsTruncated in the previous response was true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of HostedZoneId, specify the value of HostedZoneIdMarker from the previous response, which is the hosted zone ID of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

$sel:trafficPolicyInstanceNameMarker:ListTrafficPolicyInstances', listTrafficPolicyInstances_trafficPolicyInstanceNameMarker - If the value of IsTruncated in the previous response was true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancename, specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

Request Lenses

listTrafficPolicyInstances_trafficPolicyInstanceTypeMarker :: Lens' ListTrafficPolicyInstances (Maybe RRType)

If the value of IsTruncated in the previous response was true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancetype, specify the value of TrafficPolicyInstanceTypeMarker from the previous response, which is the type of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

listTrafficPolicyInstances_maxItems :: Lens' ListTrafficPolicyInstances (Maybe Text)

The maximum number of traffic policy instances that you want Amazon Route 53 to return in response to a ListTrafficPolicyInstances request. If you have more than MaxItems traffic policy instances, the value of the IsTruncated element in the response is true, and the values of HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker represent the first traffic policy instance in the next group of MaxItems traffic policy instances.

listTrafficPolicyInstances_hostedZoneIdMarker :: Lens' ListTrafficPolicyInstances (Maybe ResourceId)

If the value of IsTruncated in the previous response was true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of HostedZoneId, specify the value of HostedZoneIdMarker from the previous response, which is the hosted zone ID of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

listTrafficPolicyInstances_trafficPolicyInstanceNameMarker :: Lens' ListTrafficPolicyInstances (Maybe Text)

If the value of IsTruncated in the previous response was true, you have more traffic policy instances. To get more traffic policy instances, submit another ListTrafficPolicyInstances request. For the value of trafficpolicyinstancename, specify the value of TrafficPolicyInstanceNameMarker from the previous response, which is the name of the first traffic policy instance in the next group of traffic policy instances.

If the value of IsTruncated in the previous response was false, there are no more traffic policy instances to get.

Destructuring the Response

data ListTrafficPolicyInstancesResponse

A complex type that contains the response information for the request.

See: newListTrafficPolicyInstancesResponse smart constructor.

Constructors

ListTrafficPolicyInstancesResponse'

Fields

trafficPolicyInstanceTypeMarker :: Maybe RRType

If IsTruncated is true, TrafficPolicyInstanceTypeMarker is the DNS type of the resource record sets that are associated with the first traffic policy instance that Amazon Route 53 will return if you submit another ListTrafficPolicyInstances request.

hostedZoneIdMarker :: Maybe ResourceId

If IsTruncated is true, HostedZoneIdMarker is the ID of the hosted zone of the first traffic policy instance that Route 53 will return if you submit another ListTrafficPolicyInstances request.

trafficPolicyInstanceNameMarker :: Maybe Text

If IsTruncated is true, TrafficPolicyInstanceNameMarker is the name of the first traffic policy instance that Route 53 will return if you submit another ListTrafficPolicyInstances request.

httpStatus :: Int

The response's http status code.

trafficPolicyInstances :: [TrafficPolicyInstance]

A list that contains one TrafficPolicyInstance element for each traffic policy instance that matches the elements in the request.

isTruncated :: Bool

A flag that indicates whether there are more traffic policy instances to be listed. If the response was truncated, you can get more traffic policy instances by calling ListTrafficPolicyInstances again and specifying the values of the HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker in the corresponding request parameters.

maxItems :: Text

The value that you specified for the MaxItems parameter in the call to ListTrafficPolicyInstances that produced the current response.

newListTrafficPolicyInstancesResponse

Arguments

:: Int
$sel:httpStatus:ListTrafficPolicyInstancesResponse'

-> Bool
$sel:isTruncated:ListTrafficPolicyInstancesResponse'

-> Text
$sel:maxItems:ListTrafficPolicyInstances'

-> ListTrafficPolicyInstancesResponse

Create a value of ListTrafficPolicyInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trafficPolicyInstanceTypeMarker:ListTrafficPolicyInstances', listTrafficPolicyInstancesResponse_trafficPolicyInstanceTypeMarker - If IsTruncated is true, TrafficPolicyInstanceTypeMarker is the DNS type of the resource record sets that are associated with the first traffic policy instance that Amazon Route 53 will return if you submit another ListTrafficPolicyInstances request.

$sel:hostedZoneIdMarker:ListTrafficPolicyInstances', listTrafficPolicyInstancesResponse_hostedZoneIdMarker - If IsTruncated is true, HostedZoneIdMarker is the ID of the hosted zone of the first traffic policy instance that Route 53 will return if you submit another ListTrafficPolicyInstances request.

$sel:trafficPolicyInstanceNameMarker:ListTrafficPolicyInstances', listTrafficPolicyInstancesResponse_trafficPolicyInstanceNameMarker - If IsTruncated is true, TrafficPolicyInstanceNameMarker is the name of the first traffic policy instance that Route 53 will return if you submit another ListTrafficPolicyInstances request.

$sel:httpStatus:ListTrafficPolicyInstancesResponse', listTrafficPolicyInstancesResponse_httpStatus - The response's http status code.

$sel:trafficPolicyInstances:ListTrafficPolicyInstancesResponse', listTrafficPolicyInstancesResponse_trafficPolicyInstances - A list that contains one TrafficPolicyInstance element for each traffic policy instance that matches the elements in the request.

$sel:isTruncated:ListTrafficPolicyInstancesResponse', listTrafficPolicyInstancesResponse_isTruncated - A flag that indicates whether there are more traffic policy instances to be listed. If the response was truncated, you can get more traffic policy instances by calling ListTrafficPolicyInstances again and specifying the values of the HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker in the corresponding request parameters.

$sel:maxItems:ListTrafficPolicyInstances', listTrafficPolicyInstancesResponse_maxItems - The value that you specified for the MaxItems parameter in the call to ListTrafficPolicyInstances that produced the current response.

Response Lenses

listTrafficPolicyInstancesResponse_trafficPolicyInstanceTypeMarker :: Lens' ListTrafficPolicyInstancesResponse (Maybe RRType)

If IsTruncated is true, TrafficPolicyInstanceTypeMarker is the DNS type of the resource record sets that are associated with the first traffic policy instance that Amazon Route 53 will return if you submit another ListTrafficPolicyInstances request.

listTrafficPolicyInstancesResponse_hostedZoneIdMarker :: Lens' ListTrafficPolicyInstancesResponse (Maybe ResourceId)

If IsTruncated is true, HostedZoneIdMarker is the ID of the hosted zone of the first traffic policy instance that Route 53 will return if you submit another ListTrafficPolicyInstances request.

listTrafficPolicyInstancesResponse_trafficPolicyInstanceNameMarker :: Lens' ListTrafficPolicyInstancesResponse (Maybe Text)

If IsTruncated is true, TrafficPolicyInstanceNameMarker is the name of the first traffic policy instance that Route 53 will return if you submit another ListTrafficPolicyInstances request.

listTrafficPolicyInstancesResponse_httpStatus :: Lens' ListTrafficPolicyInstancesResponse Int

The response's http status code.

listTrafficPolicyInstancesResponse_trafficPolicyInstances :: Lens' ListTrafficPolicyInstancesResponse [TrafficPolicyInstance]

A list that contains one TrafficPolicyInstance element for each traffic policy instance that matches the elements in the request.

listTrafficPolicyInstancesResponse_isTruncated :: Lens' ListTrafficPolicyInstancesResponse Bool

A flag that indicates whether there are more traffic policy instances to be listed. If the response was truncated, you can get more traffic policy instances by calling ListTrafficPolicyInstances again and specifying the values of the HostedZoneIdMarker, TrafficPolicyInstanceNameMarker, and TrafficPolicyInstanceTypeMarker in the corresponding request parameters.

listTrafficPolicyInstancesResponse_maxItems :: Lens' ListTrafficPolicyInstancesResponse Text

The value that you specified for the MaxItems parameter in the call to ListTrafficPolicyInstances that produced the current response.