Amazonka.Route53.Types.GeoLocationDetails

Description

Documentation

data GeoLocationDetails

A complex type that contains the codes and full continent, country, and subdivision names for the specified geolocation code.

See: newGeoLocationDetails smart constructor.

Constructors

GeoLocationDetails'

Fields

subdivisionName :: Maybe Text

The full name of the subdivision. Route 53 currently supports only states in the United States.

subdivisionCode :: Maybe Text

The code for the subdivision, such as a particular state within the United States. For a list of US state abbreviations, see Appendix B: Two–Letter State and Possession Abbreviations on the United States Postal Service website. For a list of all supported subdivision codes, use the ListGeoLocations API.

countryName :: Maybe Text

The name of the country.

countryCode :: Maybe Text

The two-letter code for the country.

continentCode :: Maybe Text

The two-letter code for the continent.

continentName :: Maybe Text

The full name of the continent.

newGeoLocationDetails :: GeoLocationDetails

Create a value of GeoLocationDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subdivisionName:GeoLocationDetails', geoLocationDetails_subdivisionName - The full name of the subdivision. Route 53 currently supports only states in the United States.

$sel:subdivisionCode:GeoLocationDetails', geoLocationDetails_subdivisionCode - The code for the subdivision, such as a particular state within the United States. For a list of US state abbreviations, see Appendix B: Two–Letter State and Possession Abbreviations on the United States Postal Service website. For a list of all supported subdivision codes, use the ListGeoLocations API.

$sel:countryName:GeoLocationDetails', geoLocationDetails_countryName - The name of the country.

$sel:countryCode:GeoLocationDetails', geoLocationDetails_countryCode - The two-letter code for the country.

$sel:continentCode:GeoLocationDetails', geoLocationDetails_continentCode - The two-letter code for the continent.

$sel:continentName:GeoLocationDetails', geoLocationDetails_continentName - The full name of the continent.

geoLocationDetails_subdivisionName :: Lens' GeoLocationDetails (Maybe Text)

The full name of the subdivision. Route 53 currently supports only states in the United States.

geoLocationDetails_subdivisionCode :: Lens' GeoLocationDetails (Maybe Text)

The code for the subdivision, such as a particular state within the United States. For a list of US state abbreviations, see Appendix B: Two–Letter State and Possession Abbreviations on the United States Postal Service website. For a list of all supported subdivision codes, use the ListGeoLocations API.

geoLocationDetails_countryName :: Lens' GeoLocationDetails (Maybe Text)

The name of the country.

geoLocationDetails_countryCode :: Lens' GeoLocationDetails (Maybe Text)

The two-letter code for the country.

geoLocationDetails_continentCode :: Lens' GeoLocationDetails (Maybe Text)

The two-letter code for the continent.

geoLocationDetails_continentName :: Lens' GeoLocationDetails (Maybe Text)

The full name of the continent.