Amazonka.Route53.Types.HealthCheck

Description

Documentation

data HealthCheck

A complex type that contains information about one health check that is associated with the current Amazon Web Services account.

See: newHealthCheck smart constructor.

Constructors

HealthCheck'

Fields

linkedService :: Maybe LinkedService

If the health check was created by another service, the service that created the health check. When a health check is created by another service, you can't edit or delete it using Amazon Route 53.

cloudWatchAlarmConfiguration :: Maybe CloudWatchAlarmConfiguration

A complex type that contains information about the CloudWatch alarm that Amazon Route 53 is monitoring for this health check.

id :: Text

The identifier that Amazon Route 53 assigned to the health check when you created it. When you add or update a resource record set, you use this value to specify which health check to use. The value can be up to 64 characters long.

callerReference :: Text

A unique string that you specified when you created the health check.

healthCheckConfig :: HealthCheckConfig

A complex type that contains detailed information about one health check.

healthCheckVersion :: Natural

The version of the health check. You can optionally pass this value in a call to UpdateHealthCheck to prevent overwriting another change to the health check.

newHealthCheck

Arguments

:: Text
$sel:id:HealthCheck'

-> Text
$sel:callerReference:HealthCheck'

-> HealthCheckConfig
$sel:healthCheckConfig:HealthCheck'

-> Natural
$sel:healthCheckVersion:HealthCheck'

-> HealthCheck

Create a value of HealthCheck with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:linkedService:HealthCheck', healthCheck_linkedService - If the health check was created by another service, the service that created the health check. When a health check is created by another service, you can't edit or delete it using Amazon Route 53.

$sel:cloudWatchAlarmConfiguration:HealthCheck', healthCheck_cloudWatchAlarmConfiguration - A complex type that contains information about the CloudWatch alarm that Amazon Route 53 is monitoring for this health check.

$sel:id:HealthCheck', healthCheck_id - The identifier that Amazon Route 53 assigned to the health check when you created it. When you add or update a resource record set, you use this value to specify which health check to use. The value can be up to 64 characters long.

$sel:callerReference:HealthCheck', healthCheck_callerReference - A unique string that you specified when you created the health check.

$sel:healthCheckConfig:HealthCheck', healthCheck_healthCheckConfig - A complex type that contains detailed information about one health check.

$sel:healthCheckVersion:HealthCheck', healthCheck_healthCheckVersion - The version of the health check. You can optionally pass this value in a call to UpdateHealthCheck to prevent overwriting another change to the health check.

healthCheck_linkedService :: Lens' HealthCheck (Maybe LinkedService)

If the health check was created by another service, the service that created the health check. When a health check is created by another service, you can't edit or delete it using Amazon Route 53.

healthCheck_cloudWatchAlarmConfiguration :: Lens' HealthCheck (Maybe CloudWatchAlarmConfiguration)

A complex type that contains information about the CloudWatch alarm that Amazon Route 53 is monitoring for this health check.

healthCheck_id :: Lens' HealthCheck Text

The identifier that Amazon Route 53 assigned to the health check when you created it. When you add or update a resource record set, you use this value to specify which health check to use. The value can be up to 64 characters long.

healthCheck_callerReference :: Lens' HealthCheck Text

A unique string that you specified when you created the health check.

healthCheck_healthCheckConfig :: Lens' HealthCheck HealthCheckConfig

A complex type that contains detailed information about one health check.

healthCheck_healthCheckVersion :: Lens' HealthCheck Natural

The version of the health check. You can optionally pass this value in a call to UpdateHealthCheck to prevent overwriting another change to the health check.