Amazonka.Route53.Types.HealthCheckObservation

Description

Documentation

data HealthCheckObservation

A complex type that contains the last failure reason as reported by one Amazon Route 53 health checker.

See: newHealthCheckObservation smart constructor.

Constructors

HealthCheckObservation'

Fields

iPAddress :: Maybe Text

The IP address of the Amazon Route 53 health checker that provided the failure reason in StatusReport.

statusReport :: Maybe StatusReport

A complex type that contains the last failure reason as reported by one Amazon Route 53 health checker and the time of the failed health check.

region :: Maybe HealthCheckRegion

The region of the Amazon Route 53 health checker that provided the status in StatusReport.

newHealthCheckObservation :: HealthCheckObservation

Create a value of HealthCheckObservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iPAddress:HealthCheckObservation', healthCheckObservation_iPAddress - The IP address of the Amazon Route 53 health checker that provided the failure reason in StatusReport.

$sel:statusReport:HealthCheckObservation', healthCheckObservation_statusReport - A complex type that contains the last failure reason as reported by one Amazon Route 53 health checker and the time of the failed health check.

$sel:region:HealthCheckObservation', healthCheckObservation_region - The region of the Amazon Route 53 health checker that provided the status in StatusReport.

healthCheckObservation_iPAddress :: Lens' HealthCheckObservation (Maybe Text)

The IP address of the Amazon Route 53 health checker that provided the failure reason in StatusReport.

healthCheckObservation_statusReport :: Lens' HealthCheckObservation (Maybe StatusReport)

A complex type that contains the last failure reason as reported by one Amazon Route 53 health checker and the time of the failed health check.

healthCheckObservation_region :: Lens' HealthCheckObservation (Maybe HealthCheckRegion)

The region of the Amazon Route 53 health checker that provided the status in StatusReport.