Amazonka.Route53.Types.TrafficPolicy

Description

Documentation

data TrafficPolicy

A complex type that contains settings for a traffic policy.

See: newTrafficPolicy smart constructor.

Constructors

TrafficPolicy'

Fields

comment :: Maybe Text

The comment that you specify in the CreateTrafficPolicy request, if any.

id :: Text

The ID that Amazon Route 53 assigned to a traffic policy when you created it.

version :: Natural

The version number that Amazon Route 53 assigns to a traffic policy. For a new traffic policy, the value of Version is always 1.

name :: Text

The name that you specified when you created the traffic policy.

type' :: RRType

The DNS type of the resource record sets that Amazon Route 53 creates when you use a traffic policy to create a traffic policy instance.

document :: Text

The definition of a traffic policy in JSON format. You specify the JSON document to use for a new traffic policy in the CreateTrafficPolicy request. For more information about the JSON format, see Traffic Policy Document Format.

newTrafficPolicy

Arguments

:: Text
$sel:id:TrafficPolicy'

-> Natural
$sel:version:TrafficPolicy'

-> Text
$sel:name:TrafficPolicy'

-> RRType
$sel:type':TrafficPolicy'

-> Text
$sel:document:TrafficPolicy'

-> TrafficPolicy

Create a value of TrafficPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comment:TrafficPolicy', trafficPolicy_comment - The comment that you specify in the CreateTrafficPolicy request, if any.

$sel:id:TrafficPolicy', trafficPolicy_id - The ID that Amazon Route 53 assigned to a traffic policy when you created it.

$sel:version:TrafficPolicy', trafficPolicy_version - The version number that Amazon Route 53 assigns to a traffic policy. For a new traffic policy, the value of Version is always 1.

$sel:name:TrafficPolicy', trafficPolicy_name - The name that you specified when you created the traffic policy.

$sel:type':TrafficPolicy', trafficPolicy_type - The DNS type of the resource record sets that Amazon Route 53 creates when you use a traffic policy to create a traffic policy instance.

$sel:document:TrafficPolicy', trafficPolicy_document - The definition of a traffic policy in JSON format. You specify the JSON document to use for a new traffic policy in the CreateTrafficPolicy request. For more information about the JSON format, see Traffic Policy Document Format.

trafficPolicy_comment :: Lens' TrafficPolicy (Maybe Text)

The comment that you specify in the CreateTrafficPolicy request, if any.

trafficPolicy_id :: Lens' TrafficPolicy Text

The ID that Amazon Route 53 assigned to a traffic policy when you created it.

trafficPolicy_version :: Lens' TrafficPolicy Natural

The version number that Amazon Route 53 assigns to a traffic policy. For a new traffic policy, the value of Version is always 1.

trafficPolicy_name :: Lens' TrafficPolicy Text

The name that you specified when you created the traffic policy.

trafficPolicy_type :: Lens' TrafficPolicy RRType

The DNS type of the resource record sets that Amazon Route 53 creates when you use a traffic policy to create a traffic policy instance.

trafficPolicy_document :: Lens' TrafficPolicy Text

The definition of a traffic policy in JSON format. You specify the JSON document to use for a new traffic policy in the CreateTrafficPolicy request. For more information about the JSON format, see Traffic Policy Document Format.