Amazonka.Route53.Types.TrafficPolicyInstance

Description

Documentation

data TrafficPolicyInstance

A complex type that contains settings for the new traffic policy instance.

See: newTrafficPolicyInstance smart constructor.

Constructors

TrafficPolicyInstance'

Fields

id :: Text

The ID that Amazon Route 53 assigned to the new traffic policy instance.

hostedZoneId :: ResourceId

The ID of the hosted zone that Amazon Route 53 created resource record sets in.

name :: Text

The DNS name, such as www.example.com, for which Amazon Route 53 responds to queries by using the resource record sets that are associated with this traffic policy instance.

ttl :: Natural

The TTL that Amazon Route 53 assigned to all of the resource record sets that it created in the specified hosted zone.

state :: Text

The value of State is one of the following values:

Applied

Amazon Route 53 has finished creating resource record sets, and changes have propagated to all Route 53 edge locations.

Creating

Route 53 is creating the resource record sets. Use GetTrafficPolicyInstance to confirm that the CreateTrafficPolicyInstance request completed successfully.

Failed

Route 53 wasn't able to create or update the resource record sets. When the value of State is Failed, see Message for an explanation of what caused the request to fail.

message :: Text

If State is Failed, an explanation of the reason for the failure. If State is another value, Message is empty.

trafficPolicyId :: Text

The ID of the traffic policy that Amazon Route 53 used to create resource record sets in the specified hosted zone.

trafficPolicyVersion :: Natural

The version of the traffic policy that Amazon Route 53 used to create resource record sets in the specified hosted zone.

trafficPolicyType :: RRType

The DNS type that Amazon Route 53 assigned to all of the resource record sets that it created for this traffic policy instance.

newTrafficPolicyInstance

Arguments

:: Text
$sel:id:TrafficPolicyInstance'

-> ResourceId
$sel:hostedZoneId:TrafficPolicyInstance'

-> Text
$sel:name:TrafficPolicyInstance'

-> Natural
$sel:ttl:TrafficPolicyInstance'

-> Text
$sel:state:TrafficPolicyInstance'

-> Text
$sel:message:TrafficPolicyInstance'

-> Text
$sel:trafficPolicyId:TrafficPolicyInstance'

-> Natural
$sel:trafficPolicyVersion:TrafficPolicyInstance'

-> RRType
$sel:trafficPolicyType:TrafficPolicyInstance'

-> TrafficPolicyInstance

Create a value of TrafficPolicyInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:TrafficPolicyInstance', trafficPolicyInstance_id - The ID that Amazon Route 53 assigned to the new traffic policy instance.

$sel:hostedZoneId:TrafficPolicyInstance', trafficPolicyInstance_hostedZoneId - The ID of the hosted zone that Amazon Route 53 created resource record sets in.

$sel:name:TrafficPolicyInstance', trafficPolicyInstance_name - The DNS name, such as www.example.com, for which Amazon Route 53 responds to queries by using the resource record sets that are associated with this traffic policy instance.

$sel:ttl:TrafficPolicyInstance', trafficPolicyInstance_ttl - The TTL that Amazon Route 53 assigned to all of the resource record sets that it created in the specified hosted zone.

$sel:state:TrafficPolicyInstance', trafficPolicyInstance_state - The value of State is one of the following values:

Applied

Amazon Route 53 has finished creating resource record sets, and changes have propagated to all Route 53 edge locations.

Creating

Route 53 is creating the resource record sets. Use GetTrafficPolicyInstance to confirm that the CreateTrafficPolicyInstance request completed successfully.

Failed

Route 53 wasn't able to create or update the resource record sets. When the value of State is Failed, see Message for an explanation of what caused the request to fail.

$sel:message:TrafficPolicyInstance', trafficPolicyInstance_message - If State is Failed, an explanation of the reason for the failure. If State is another value, Message is empty.

$sel:trafficPolicyId:TrafficPolicyInstance', trafficPolicyInstance_trafficPolicyId - The ID of the traffic policy that Amazon Route 53 used to create resource record sets in the specified hosted zone.

$sel:trafficPolicyVersion:TrafficPolicyInstance', trafficPolicyInstance_trafficPolicyVersion - The version of the traffic policy that Amazon Route 53 used to create resource record sets in the specified hosted zone.

$sel:trafficPolicyType:TrafficPolicyInstance', trafficPolicyInstance_trafficPolicyType - The DNS type that Amazon Route 53 assigned to all of the resource record sets that it created for this traffic policy instance.

trafficPolicyInstance_id :: Lens' TrafficPolicyInstance Text

The ID that Amazon Route 53 assigned to the new traffic policy instance.

trafficPolicyInstance_hostedZoneId :: Lens' TrafficPolicyInstance ResourceId

The ID of the hosted zone that Amazon Route 53 created resource record sets in.

trafficPolicyInstance_name :: Lens' TrafficPolicyInstance Text

The DNS name, such as www.example.com, for which Amazon Route 53 responds to queries by using the resource record sets that are associated with this traffic policy instance.

trafficPolicyInstance_ttl :: Lens' TrafficPolicyInstance Natural

The TTL that Amazon Route 53 assigned to all of the resource record sets that it created in the specified hosted zone.

trafficPolicyInstance_state :: Lens' TrafficPolicyInstance Text

The value of State is one of the following values:

Applied

Amazon Route 53 has finished creating resource record sets, and changes have propagated to all Route 53 edge locations.

Creating

Route 53 is creating the resource record sets. Use GetTrafficPolicyInstance to confirm that the CreateTrafficPolicyInstance request completed successfully.

Failed

Route 53 wasn't able to create or update the resource record sets. When the value of State is Failed, see Message for an explanation of what caused the request to fail.

trafficPolicyInstance_message :: Lens' TrafficPolicyInstance Text

If State is Failed, an explanation of the reason for the failure. If State is another value, Message is empty.

trafficPolicyInstance_trafficPolicyId :: Lens' TrafficPolicyInstance Text

The ID of the traffic policy that Amazon Route 53 used to create resource record sets in the specified hosted zone.

trafficPolicyInstance_trafficPolicyVersion :: Lens' TrafficPolicyInstance Natural

The version of the traffic policy that Amazon Route 53 used to create resource record sets in the specified hosted zone.

trafficPolicyInstance_trafficPolicyType :: Lens' TrafficPolicyInstance RRType

The DNS type that Amazon Route 53 assigned to all of the resource record sets that it created for this traffic policy instance.