Amazonka.Route53.UpdateTrafficPolicyInstance

Description

Updates the resource record sets in a specified hosted zone that were created based on the settings in a specified traffic policy version.

When you update a traffic policy instance, Amazon Route 53 continues to respond to DNS queries for the root resource record set name (such as example.com) while it replaces one group of resource record sets with another. Route 53 performs the following operations:

Route 53 creates a new group of resource record sets based on the specified traffic policy. This is true regardless of how significant the differences are between the existing resource record sets and the new resource record sets.

When all of the new resource record sets have been created, Route 53 starts to respond to DNS queries for the root resource record set name (such as example.com) by using the new resource record sets.

Route 53 deletes the old group of resource record sets that are associated with the root resource record set name.

Creating a Request

data UpdateTrafficPolicyInstance

A complex type that contains information about the resource record sets that you want to update based on a specified traffic policy instance.

See: newUpdateTrafficPolicyInstance smart constructor.

Constructors

UpdateTrafficPolicyInstance'

Fields

id :: Text

The ID of the traffic policy instance that you want to update.

ttl :: Natural

The TTL that you want Amazon Route 53 to assign to all of the updated resource record sets.

trafficPolicyId :: Text

The ID of the traffic policy that you want Amazon Route 53 to use to update resource record sets for the specified traffic policy instance.

trafficPolicyVersion :: Natural

The version of the traffic policy that you want Amazon Route 53 to use to update resource record sets for the specified traffic policy instance.

newUpdateTrafficPolicyInstance

Arguments

:: Text
$sel:id:UpdateTrafficPolicyInstance'

-> Natural
$sel:ttl:UpdateTrafficPolicyInstance'

-> Text
$sel:trafficPolicyId:UpdateTrafficPolicyInstance'

-> Natural
$sel:trafficPolicyVersion:UpdateTrafficPolicyInstance'

-> UpdateTrafficPolicyInstance

Create a value of UpdateTrafficPolicyInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:UpdateTrafficPolicyInstance', updateTrafficPolicyInstance_id - The ID of the traffic policy instance that you want to update.

$sel:ttl:UpdateTrafficPolicyInstance', updateTrafficPolicyInstance_ttl - The TTL that you want Amazon Route 53 to assign to all of the updated resource record sets.

$sel:trafficPolicyId:UpdateTrafficPolicyInstance', updateTrafficPolicyInstance_trafficPolicyId - The ID of the traffic policy that you want Amazon Route 53 to use to update resource record sets for the specified traffic policy instance.

$sel:trafficPolicyVersion:UpdateTrafficPolicyInstance', updateTrafficPolicyInstance_trafficPolicyVersion - The version of the traffic policy that you want Amazon Route 53 to use to update resource record sets for the specified traffic policy instance.

Request Lenses

updateTrafficPolicyInstance_id :: Lens' UpdateTrafficPolicyInstance Text

The ID of the traffic policy instance that you want to update.

updateTrafficPolicyInstance_ttl :: Lens' UpdateTrafficPolicyInstance Natural

The TTL that you want Amazon Route 53 to assign to all of the updated resource record sets.

updateTrafficPolicyInstance_trafficPolicyId :: Lens' UpdateTrafficPolicyInstance Text

The ID of the traffic policy that you want Amazon Route 53 to use to update resource record sets for the specified traffic policy instance.

updateTrafficPolicyInstance_trafficPolicyVersion :: Lens' UpdateTrafficPolicyInstance Natural

The version of the traffic policy that you want Amazon Route 53 to use to update resource record sets for the specified traffic policy instance.

Destructuring the Response

data UpdateTrafficPolicyInstanceResponse

A complex type that contains information about the resource record sets that Amazon Route 53 created based on a specified traffic policy.

See: newUpdateTrafficPolicyInstanceResponse smart constructor.

Constructors

UpdateTrafficPolicyInstanceResponse'

Fields

httpStatus :: Int

The response's http status code.

trafficPolicyInstance :: TrafficPolicyInstance

A complex type that contains settings for the updated traffic policy instance.

newUpdateTrafficPolicyInstanceResponse

Arguments

:: Int
$sel:httpStatus:UpdateTrafficPolicyInstanceResponse'

-> TrafficPolicyInstance
$sel:trafficPolicyInstance:UpdateTrafficPolicyInstanceResponse'

-> UpdateTrafficPolicyInstanceResponse

Create a value of UpdateTrafficPolicyInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateTrafficPolicyInstanceResponse', updateTrafficPolicyInstanceResponse_httpStatus - The response's http status code.

$sel:trafficPolicyInstance:UpdateTrafficPolicyInstanceResponse', updateTrafficPolicyInstanceResponse_trafficPolicyInstance - A complex type that contains settings for the updated traffic policy instance.

Response Lenses

updateTrafficPolicyInstanceResponse_httpStatus :: Lens' UpdateTrafficPolicyInstanceResponse Int

The response's http status code.

updateTrafficPolicyInstanceResponse_trafficPolicyInstance :: Lens' UpdateTrafficPolicyInstanceResponse TrafficPolicyInstance

A complex type that contains settings for the updated traffic policy instance.