Amazonka.Route53AutoNaming.GetInstancesHealthStatus

Description

Gets the current health status (Healthy, Unhealthy, or Unknown) of one or more instances that are associated with a specified service.

There's a brief delay between when you register an instance and when the health status for the instance is available.

Creating a Request

data GetInstancesHealthStatus

See: newGetInstancesHealthStatus smart constructor.

Constructors

GetInstancesHealthStatus'

Fields

nextToken :: Maybe Text

For the first GetInstancesHealthStatus request, omit this value.

If more than MaxResults instances match the specified criteria, you can submit another GetInstancesHealthStatus request to get the next group of results. Specify the value of NextToken from the previous response in the next request.

instances :: Maybe (NonEmpty Text)

An array that contains the IDs of all the instances that you want to get the health status for.

If you omit Instances, Cloud Map returns the health status for all the instances that are associated with the specified service.

To get the IDs for the instances that you've registered by using a specified service, submit a ListInstances request.

maxResults :: Maybe Natural

The maximum number of instances that you want Cloud Map to return in the response to a GetInstancesHealthStatus request. If you don't specify a value for MaxResults, Cloud Map returns up to 100 instances.

serviceId :: Text

The ID of the service that the instance is associated with.

newGetInstancesHealthStatus

Arguments

:: Text
$sel:serviceId:GetInstancesHealthStatus'

-> GetInstancesHealthStatus

Create a value of GetInstancesHealthStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetInstancesHealthStatus', getInstancesHealthStatus_nextToken - For the first GetInstancesHealthStatus request, omit this value.

If more than MaxResults instances match the specified criteria, you can submit another GetInstancesHealthStatus request to get the next group of results. Specify the value of NextToken from the previous response in the next request.

$sel:instances:GetInstancesHealthStatus', getInstancesHealthStatus_instances - An array that contains the IDs of all the instances that you want to get the health status for.

If you omit Instances, Cloud Map returns the health status for all the instances that are associated with the specified service.

To get the IDs for the instances that you've registered by using a specified service, submit a ListInstances request.

$sel:maxResults:GetInstancesHealthStatus', getInstancesHealthStatus_maxResults - The maximum number of instances that you want Cloud Map to return in the response to a GetInstancesHealthStatus request. If you don't specify a value for MaxResults, Cloud Map returns up to 100 instances.

$sel:serviceId:GetInstancesHealthStatus', getInstancesHealthStatus_serviceId - The ID of the service that the instance is associated with.

Request Lenses

getInstancesHealthStatus_nextToken :: Lens' GetInstancesHealthStatus (Maybe Text)

For the first GetInstancesHealthStatus request, omit this value.

If more than MaxResults instances match the specified criteria, you can submit another GetInstancesHealthStatus request to get the next group of results. Specify the value of NextToken from the previous response in the next request.

getInstancesHealthStatus_instances :: Lens' GetInstancesHealthStatus (Maybe (NonEmpty Text))

An array that contains the IDs of all the instances that you want to get the health status for.

If you omit Instances, Cloud Map returns the health status for all the instances that are associated with the specified service.

To get the IDs for the instances that you've registered by using a specified service, submit a ListInstances request.

getInstancesHealthStatus_maxResults :: Lens' GetInstancesHealthStatus (Maybe Natural)

The maximum number of instances that you want Cloud Map to return in the response to a GetInstancesHealthStatus request. If you don't specify a value for MaxResults, Cloud Map returns up to 100 instances.

getInstancesHealthStatus_serviceId :: Lens' GetInstancesHealthStatus Text

The ID of the service that the instance is associated with.

Destructuring the Response

data GetInstancesHealthStatusResponse

See: newGetInstancesHealthStatusResponse smart constructor.

Constructors

GetInstancesHealthStatusResponse'

Fields

status :: Maybe (HashMap Text HealthStatus)

A complex type that contains the IDs and the health status of the instances that you specified in the GetInstancesHealthStatus request.

nextToken :: Maybe Text

If more than MaxResults instances match the specified criteria, you can submit another GetInstancesHealthStatus request to get the next group of results. Specify the value of NextToken from the previous response in the next request.

httpStatus :: Int

The response's http status code.

newGetInstancesHealthStatusResponse

Arguments

:: Int
$sel:httpStatus:GetInstancesHealthStatusResponse'

-> GetInstancesHealthStatusResponse

Create a value of GetInstancesHealthStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetInstancesHealthStatusResponse', getInstancesHealthStatusResponse_status - A complex type that contains the IDs and the health status of the instances that you specified in the GetInstancesHealthStatus request.

$sel:nextToken:GetInstancesHealthStatus', getInstancesHealthStatusResponse_nextToken - If more than MaxResults instances match the specified criteria, you can submit another GetInstancesHealthStatus request to get the next group of results. Specify the value of NextToken from the previous response in the next request.

$sel:httpStatus:GetInstancesHealthStatusResponse', getInstancesHealthStatusResponse_httpStatus - The response's http status code.

Response Lenses

getInstancesHealthStatusResponse_status :: Lens' GetInstancesHealthStatusResponse (Maybe (HashMap Text HealthStatus))

A complex type that contains the IDs and the health status of the instances that you specified in the GetInstancesHealthStatus request.

getInstancesHealthStatusResponse_nextToken :: Lens' GetInstancesHealthStatusResponse (Maybe Text)

If more than MaxResults instances match the specified criteria, you can submit another GetInstancesHealthStatus request to get the next group of results. Specify the value of NextToken from the previous response in the next request.

getInstancesHealthStatusResponse_httpStatus :: Lens' GetInstancesHealthStatusResponse Int

The response's http status code.