Amazonka.Route53AutoNaming.ListOperations

Description

Lists operations that match the criteria that you specify.

This operation returns paginated results.

Creating a Request

data ListOperations

See: newListOperations smart constructor.

Constructors

ListOperations'

Fields

filters :: Maybe [OperationFilter]

A complex type that contains specifications for the operations that you want to list, for example, operations that you started between a specified start date and end date.

If you specify more than one filter, an operation must match all filters to be returned by ListOperations.

nextToken :: Maybe Text

For the first ListOperations request, omit this value.

If the response contains NextToken, submit another ListOperations request to get the next group of results. Specify the value of NextToken from the previous response in the next request.

Cloud Map gets MaxResults operations and then filters them based on the specified criteria. It's possible that no operations in the first MaxResults operations matched the specified criteria but that subsequent groups of MaxResults operations do contain operations that match the criteria.

maxResults :: Maybe Natural

The maximum number of items that you want Cloud Map to return in the response to a ListOperations request. If you don't specify a value for MaxResults, Cloud Map returns up to 100 operations.

newListOperations :: ListOperations

Create a value of ListOperations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListOperations', listOperations_filters - A complex type that contains specifications for the operations that you want to list, for example, operations that you started between a specified start date and end date.

If you specify more than one filter, an operation must match all filters to be returned by ListOperations.

$sel:nextToken:ListOperations', listOperations_nextToken - For the first ListOperations request, omit this value.

If the response contains NextToken, submit another ListOperations request to get the next group of results. Specify the value of NextToken from the previous response in the next request.

Cloud Map gets MaxResults operations and then filters them based on the specified criteria. It's possible that no operations in the first MaxResults operations matched the specified criteria but that subsequent groups of MaxResults operations do contain operations that match the criteria.

$sel:maxResults:ListOperations', listOperations_maxResults - The maximum number of items that you want Cloud Map to return in the response to a ListOperations request. If you don't specify a value for MaxResults, Cloud Map returns up to 100 operations.

Request Lenses

listOperations_filters :: Lens' ListOperations (Maybe [OperationFilter])

A complex type that contains specifications for the operations that you want to list, for example, operations that you started between a specified start date and end date.

If you specify more than one filter, an operation must match all filters to be returned by ListOperations.

listOperations_nextToken :: Lens' ListOperations (Maybe Text)

For the first ListOperations request, omit this value.

If the response contains NextToken, submit another ListOperations request to get the next group of results. Specify the value of NextToken from the previous response in the next request.

Cloud Map gets MaxResults operations and then filters them based on the specified criteria. It's possible that no operations in the first MaxResults operations matched the specified criteria but that subsequent groups of MaxResults operations do contain operations that match the criteria.

listOperations_maxResults :: Lens' ListOperations (Maybe Natural)

The maximum number of items that you want Cloud Map to return in the response to a ListOperations request. If you don't specify a value for MaxResults, Cloud Map returns up to 100 operations.

Destructuring the Response

data ListOperationsResponse

See: newListOperationsResponse smart constructor.

Constructors

ListOperationsResponse'

Fields

nextToken :: Maybe Text

If the response contains NextToken, submit another ListOperations request to get the next group of results. Specify the value of NextToken from the previous response in the next request.

Cloud Map gets MaxResults operations and then filters them based on the specified criteria. It's possible that no operations in the first MaxResults operations matched the specified criteria but that subsequent groups of MaxResults operations do contain operations that match the criteria.

operations :: Maybe [OperationSummary]

Summary information about the operations that match the specified criteria.

httpStatus :: Int

The response's http status code.

newListOperationsResponse

Arguments

:: Int
$sel:httpStatus:ListOperationsResponse'

-> ListOperationsResponse

Create a value of ListOperationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListOperations', listOperationsResponse_nextToken - If the response contains NextToken, submit another ListOperations request to get the next group of results. Specify the value of NextToken from the previous response in the next request.

Cloud Map gets MaxResults operations and then filters them based on the specified criteria. It's possible that no operations in the first MaxResults operations matched the specified criteria but that subsequent groups of MaxResults operations do contain operations that match the criteria.

$sel:operations:ListOperationsResponse', listOperationsResponse_operations - Summary information about the operations that match the specified criteria.

$sel:httpStatus:ListOperationsResponse', listOperationsResponse_httpStatus - The response's http status code.

Response Lenses

listOperationsResponse_nextToken :: Lens' ListOperationsResponse (Maybe Text)

If the response contains NextToken, submit another ListOperations request to get the next group of results. Specify the value of NextToken from the previous response in the next request.

Cloud Map gets MaxResults operations and then filters them based on the specified criteria. It's possible that no operations in the first MaxResults operations matched the specified criteria but that subsequent groups of MaxResults operations do contain operations that match the criteria.

listOperationsResponse_operations :: Lens' ListOperationsResponse (Maybe [OperationSummary])

Summary information about the operations that match the specified criteria.

listOperationsResponse_httpStatus :: Lens' ListOperationsResponse Int

The response's http status code.