Amazonka.Route53AutoNaming.Types.DnsProperties

Description

Documentation

data DnsProperties

A complex type that contains the ID for the Route 53 hosted zone that Cloud Map creates when you create a namespace.

See: newDnsProperties smart constructor.

Constructors

DnsProperties'

Fields

hostedZoneId :: Maybe Text

The ID for the Route 53 hosted zone that Cloud Map creates when you create a namespace.

soa :: Maybe SOA

Start of Authority (SOA) record for the hosted zone.

newDnsProperties :: DnsProperties

Create a value of DnsProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostedZoneId:DnsProperties', dnsProperties_hostedZoneId - The ID for the Route 53 hosted zone that Cloud Map creates when you create a namespace.

$sel:soa:DnsProperties', dnsProperties_soa - Start of Authority (SOA) record for the hosted zone.

dnsProperties_hostedZoneId :: Lens' DnsProperties (Maybe Text)

The ID for the Route 53 hosted zone that Cloud Map creates when you create a namespace.

dnsProperties_soa :: Lens' DnsProperties (Maybe SOA)

Start of Authority (SOA) record for the hosted zone.