Amazonka.Route53AutoNaming.Types.Operation

Description

Documentation

data Operation

A complex type that contains information about a specified operation.

See: newOperation smart constructor.

Constructors

Operation'

Fields

status :: Maybe OperationStatus

The status of the operation. Values include the following:

SUBMITTED

This is the initial state that occurs immediately after you submit a request.

PENDING

Cloud Map is performing the operation.

SUCCESS

The operation succeeded.

FAIL

The operation failed. For the failure reason, see ErrorMessage.

updateDate :: Maybe POSIX

The date and time that the value of Status changed to the current value, in Unix date/time format and Coordinated Universal Time (UTC). The value of UpdateDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

createDate :: Maybe POSIX

The date and time that the request was submitted, in Unix date/time format and Coordinated Universal Time (UTC). The value of CreateDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

targets :: Maybe (HashMap OperationTargetType Text)

The name of the target entity that's associated with the operation:

NAMESPACE

The namespace ID is returned in the ResourceId property.

SERVICE

The service ID is returned in the ResourceId property.

INSTANCE

The instance ID is returned in the ResourceId property.

errorCode :: Maybe Text

The code associated with ErrorMessage. Values for ErrorCode include the following:

ACCESS_DENIED

CANNOT_CREATE_HOSTED_ZONE

EXPIRED_TOKEN

HOSTED_ZONE_NOT_FOUND

INTERNAL_FAILURE

INVALID_CHANGE_BATCH

THROTTLED_REQUEST

id :: Maybe Text

The ID of the operation that you want to get information about.

type' :: Maybe OperationType

The name of the operation that's associated with the specified ID.

errorMessage :: Maybe Text

If the value of Status is FAIL, the reason that the operation failed.

newOperation :: Operation

Create a value of Operation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Operation', operation_status - The status of the operation. Values include the following:

SUBMITTED

This is the initial state that occurs immediately after you submit a request.

PENDING

Cloud Map is performing the operation.

SUCCESS

The operation succeeded.

FAIL

The operation failed. For the failure reason, see ErrorMessage.

$sel:updateDate:Operation', operation_updateDate - The date and time that the value of Status changed to the current value, in Unix date/time format and Coordinated Universal Time (UTC). The value of UpdateDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:createDate:Operation', operation_createDate - The date and time that the request was submitted, in Unix date/time format and Coordinated Universal Time (UTC). The value of CreateDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:targets:Operation', operation_targets - The name of the target entity that's associated with the operation:

NAMESPACE

The namespace ID is returned in the ResourceId property.

SERVICE

The service ID is returned in the ResourceId property.

INSTANCE

The instance ID is returned in the ResourceId property.

$sel:errorCode:Operation', operation_errorCode - The code associated with ErrorMessage. Values for ErrorCode include the following:

ACCESS_DENIED

CANNOT_CREATE_HOSTED_ZONE

EXPIRED_TOKEN

HOSTED_ZONE_NOT_FOUND

INTERNAL_FAILURE

INVALID_CHANGE_BATCH

THROTTLED_REQUEST

$sel:id:Operation', operation_id - The ID of the operation that you want to get information about.

$sel:type':Operation', operation_type - The name of the operation that's associated with the specified ID.

$sel:errorMessage:Operation', operation_errorMessage - If the value of Status is FAIL, the reason that the operation failed.

operation_status :: Lens' Operation (Maybe OperationStatus)

The status of the operation. Values include the following:

SUBMITTED

This is the initial state that occurs immediately after you submit a request.

PENDING

Cloud Map is performing the operation.

SUCCESS

The operation succeeded.

FAIL

The operation failed. For the failure reason, see ErrorMessage.

operation_updateDate :: Lens' Operation (Maybe UTCTime)

The date and time that the value of Status changed to the current value, in Unix date/time format and Coordinated Universal Time (UTC). The value of UpdateDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

operation_createDate :: Lens' Operation (Maybe UTCTime)

The date and time that the request was submitted, in Unix date/time format and Coordinated Universal Time (UTC). The value of CreateDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

operation_targets :: Lens' Operation (Maybe (HashMap OperationTargetType Text))

The name of the target entity that's associated with the operation:

NAMESPACE

The namespace ID is returned in the ResourceId property.

SERVICE

The service ID is returned in the ResourceId property.

INSTANCE

The instance ID is returned in the ResourceId property.

operation_errorCode :: Lens' Operation (Maybe Text)

The code associated with ErrorMessage. Values for ErrorCode include the following:

ACCESS_DENIED

CANNOT_CREATE_HOSTED_ZONE

EXPIRED_TOKEN

HOSTED_ZONE_NOT_FOUND

INTERNAL_FAILURE

INVALID_CHANGE_BATCH

THROTTLED_REQUEST

operation_id :: Lens' Operation (Maybe Text)

The ID of the operation that you want to get information about.

operation_type :: Lens' Operation (Maybe OperationType)

The name of the operation that's associated with the specified ID.

operation_errorMessage :: Lens' Operation (Maybe Text)

If the value of Status is FAIL, the reason that the operation failed.