Amazonka.Route53AutoNaming.UpdateHttpNamespace

Description

Updates an HTTP namespace.

Creating a Request

data UpdateHttpNamespace

See: newUpdateHttpNamespace smart constructor.

Constructors

UpdateHttpNamespace'

Fields

updaterRequestId :: Maybe Text

A unique string that identifies the request and that allows failed UpdateHttpNamespace requests to be retried without the risk of running the operation twice. UpdaterRequestId can be any unique string (for example, a date/timestamp).

id :: Text

The ID of the namespace that you want to update.

namespace :: HttpNamespaceChange

Updated properties for the the HTTP namespace.

newUpdateHttpNamespace

Arguments

:: Text
$sel:id:UpdateHttpNamespace'

-> HttpNamespaceChange
$sel:namespace:UpdateHttpNamespace'

-> UpdateHttpNamespace

Create a value of UpdateHttpNamespace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updaterRequestId:UpdateHttpNamespace', updateHttpNamespace_updaterRequestId - A unique string that identifies the request and that allows failed UpdateHttpNamespace requests to be retried without the risk of running the operation twice. UpdaterRequestId can be any unique string (for example, a date/timestamp).

$sel:id:UpdateHttpNamespace', updateHttpNamespace_id - The ID of the namespace that you want to update.

$sel:namespace:UpdateHttpNamespace', updateHttpNamespace_namespace - Updated properties for the the HTTP namespace.

Request Lenses

updateHttpNamespace_updaterRequestId :: Lens' UpdateHttpNamespace (Maybe Text)

A unique string that identifies the request and that allows failed UpdateHttpNamespace requests to be retried without the risk of running the operation twice. UpdaterRequestId can be any unique string (for example, a date/timestamp).

updateHttpNamespace_id :: Lens' UpdateHttpNamespace Text

The ID of the namespace that you want to update.

updateHttpNamespace_namespace :: Lens' UpdateHttpNamespace HttpNamespaceChange

Updated properties for the the HTTP namespace.

Destructuring the Response

data UpdateHttpNamespaceResponse

See: newUpdateHttpNamespaceResponse smart constructor.

Constructors

UpdateHttpNamespaceResponse'

Fields

operationId :: Maybe Text

A value that you can use to determine whether the request completed successfully. To get the status of the operation, see GetOperation.

httpStatus :: Int

The response's http status code.

newUpdateHttpNamespaceResponse

Arguments

:: Int
$sel:httpStatus:UpdateHttpNamespaceResponse'

-> UpdateHttpNamespaceResponse

Create a value of UpdateHttpNamespaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:UpdateHttpNamespaceResponse', updateHttpNamespaceResponse_operationId - A value that you can use to determine whether the request completed successfully. To get the status of the operation, see GetOperation.

$sel:httpStatus:UpdateHttpNamespaceResponse', updateHttpNamespaceResponse_httpStatus - The response's http status code.

Response Lenses

updateHttpNamespaceResponse_operationId :: Lens' UpdateHttpNamespaceResponse (Maybe Text)

A value that you can use to determine whether the request completed successfully. To get the status of the operation, see GetOperation.

updateHttpNamespaceResponse_httpStatus :: Lens' UpdateHttpNamespaceResponse Int

The response's http status code.