Amazonka.Route53Domains.AcceptDomainTransferFromAnotherAwsAccount

Description

Accepts the transfer of a domain from another AWS account to the current AWS account. You initiate a transfer between AWS accounts using TransferDomainToAnotherAwsAccount.

Use either ListOperations or GetOperationDetail to determine whether the operation succeeded. GetOperationDetail provides additional information, for example, Domain Transfer from Aws Account 111122223333 has been cancelled.

Creating a Request

data AcceptDomainTransferFromAnotherAwsAccount

The AcceptDomainTransferFromAnotherAwsAccount request includes the following elements.

See: newAcceptDomainTransferFromAnotherAwsAccount smart constructor.

Constructors

AcceptDomainTransferFromAnotherAwsAccount'

Fields

domainName :: Text

The name of the domain that was specified when another AWS account submitted a TransferDomainToAnotherAwsAccount request.

password :: Text

The password that was returned by the TransferDomainToAnotherAwsAccount request.

newAcceptDomainTransferFromAnotherAwsAccount

Arguments

:: Text
$sel:domainName:AcceptDomainTransferFromAnotherAwsAccount'

-> Text
$sel:password:AcceptDomainTransferFromAnotherAwsAccount'

-> AcceptDomainTransferFromAnotherAwsAccount

Create a value of AcceptDomainTransferFromAnotherAwsAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:AcceptDomainTransferFromAnotherAwsAccount', acceptDomainTransferFromAnotherAwsAccount_domainName - The name of the domain that was specified when another AWS account submitted a TransferDomainToAnotherAwsAccount request.

$sel:password:AcceptDomainTransferFromAnotherAwsAccount', acceptDomainTransferFromAnotherAwsAccount_password - The password that was returned by the TransferDomainToAnotherAwsAccount request.

Request Lenses

acceptDomainTransferFromAnotherAwsAccount_domainName :: Lens' AcceptDomainTransferFromAnotherAwsAccount Text

The name of the domain that was specified when another AWS account submitted a TransferDomainToAnotherAwsAccount request.

acceptDomainTransferFromAnotherAwsAccount_password :: Lens' AcceptDomainTransferFromAnotherAwsAccount Text

The password that was returned by the TransferDomainToAnotherAwsAccount request.

Destructuring the Response

data AcceptDomainTransferFromAnotherAwsAccountResponse

The AcceptDomainTransferFromAnotherAwsAccount response includes the following element.

See: newAcceptDomainTransferFromAnotherAwsAccountResponse smart constructor.

Constructors

AcceptDomainTransferFromAnotherAwsAccountResponse'

Fields

operationId :: Maybe Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

httpStatus :: Int

The response's http status code.

newAcceptDomainTransferFromAnotherAwsAccountResponse

Arguments

:: Int
$sel:httpStatus:AcceptDomainTransferFromAnotherAwsAccountResponse'

-> AcceptDomainTransferFromAnotherAwsAccountResponse

Create a value of AcceptDomainTransferFromAnotherAwsAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:AcceptDomainTransferFromAnotherAwsAccountResponse', acceptDomainTransferFromAnotherAwsAccountResponse_operationId - Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

$sel:httpStatus:AcceptDomainTransferFromAnotherAwsAccountResponse', acceptDomainTransferFromAnotherAwsAccountResponse_httpStatus - The response's http status code.

Response Lenses

acceptDomainTransferFromAnotherAwsAccountResponse_operationId :: Lens' AcceptDomainTransferFromAnotherAwsAccountResponse (Maybe Text)

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

acceptDomainTransferFromAnotherAwsAccountResponse_httpStatus :: Lens' AcceptDomainTransferFromAnotherAwsAccountResponse Int

The response's http status code.