Amazonka.Route53Domains.RejectDomainTransferFromAnotherAwsAccount

Description

Rejects the transfer of a domain from another AWS account to the current AWS account. You initiate a transfer between AWS accounts using TransferDomainToAnotherAwsAccount.

Use either ListOperations or GetOperationDetail to determine whether the operation succeeded. GetOperationDetail provides additional information, for example, Domain Transfer from Aws Account 111122223333 has been cancelled.

Creating a Request

data RejectDomainTransferFromAnotherAwsAccount

The RejectDomainTransferFromAnotherAwsAccount request includes the following element.

See: newRejectDomainTransferFromAnotherAwsAccount smart constructor.

Constructors

RejectDomainTransferFromAnotherAwsAccount'

Fields

domainName :: Text

The name of the domain that was specified when another AWS account submitted a TransferDomainToAnotherAwsAccount request.

newRejectDomainTransferFromAnotherAwsAccount

Arguments

:: Text
$sel:domainName:RejectDomainTransferFromAnotherAwsAccount'

-> RejectDomainTransferFromAnotherAwsAccount

Create a value of RejectDomainTransferFromAnotherAwsAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:RejectDomainTransferFromAnotherAwsAccount', rejectDomainTransferFromAnotherAwsAccount_domainName - The name of the domain that was specified when another AWS account submitted a TransferDomainToAnotherAwsAccount request.

Request Lenses

rejectDomainTransferFromAnotherAwsAccount_domainName :: Lens' RejectDomainTransferFromAnotherAwsAccount Text

The name of the domain that was specified when another AWS account submitted a TransferDomainToAnotherAwsAccount request.

Destructuring the Response

data RejectDomainTransferFromAnotherAwsAccountResponse

The RejectDomainTransferFromAnotherAwsAccount response includes the following element.

See: newRejectDomainTransferFromAnotherAwsAccountResponse smart constructor.

Constructors

RejectDomainTransferFromAnotherAwsAccountResponse'

Fields

operationId :: Maybe Text

The identifier that TransferDomainToAnotherAwsAccount returned to track the progress of the request. Because the transfer request was rejected, the value is no longer valid, and you can't use GetOperationDetail to query the operation status.

httpStatus :: Int

The response's http status code.

newRejectDomainTransferFromAnotherAwsAccountResponse

Arguments

:: Int
$sel:httpStatus:RejectDomainTransferFromAnotherAwsAccountResponse'

-> RejectDomainTransferFromAnotherAwsAccountResponse

Create a value of RejectDomainTransferFromAnotherAwsAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:RejectDomainTransferFromAnotherAwsAccountResponse', rejectDomainTransferFromAnotherAwsAccountResponse_operationId - The identifier that TransferDomainToAnotherAwsAccount returned to track the progress of the request. Because the transfer request was rejected, the value is no longer valid, and you can't use GetOperationDetail to query the operation status.

$sel:httpStatus:RejectDomainTransferFromAnotherAwsAccountResponse', rejectDomainTransferFromAnotherAwsAccountResponse_httpStatus - The response's http status code.

Response Lenses

rejectDomainTransferFromAnotherAwsAccountResponse_operationId :: Lens' RejectDomainTransferFromAnotherAwsAccountResponse (Maybe Text)

The identifier that TransferDomainToAnotherAwsAccount returned to track the progress of the request. Because the transfer request was rejected, the value is no longer valid, and you can't use GetOperationDetail to query the operation status.

rejectDomainTransferFromAnotherAwsAccountResponse_httpStatus :: Lens' RejectDomainTransferFromAnotherAwsAccountResponse Int

The response's http status code.