Amazonka.Route53Domains.TransferDomain

Description

Transfers a domain from another registrar to Amazon Route 53. When the transfer is complete, the domain is registered either with Amazon Registrar (for .com, .net, and .org domains) or with our registrar associate, Gandi (for all other TLDs).

For more information about transferring domains, see the following topics:

For transfer requirements, a detailed procedure, and information about viewing the status of a domain that you're transferring to Route 53, see Transferring Registration for a Domain to Amazon Route 53 in the Amazon Route 53 Developer Guide.

For information about how to transfer a domain from one AWS account to another, see TransferDomainToAnotherAwsAccount.

For information about how to transfer a domain to another domain registrar, see Transferring a Domain from Amazon Route 53 to Another Registrar in the Amazon Route 53 Developer Guide.

If the registrar for your domain is also the DNS service provider for the domain, we highly recommend that you transfer your DNS service to Route 53 or to another DNS service provider before you transfer your registration. Some registrars provide free DNS service when you purchase a domain registration. When you transfer the registration, the previous registrar will not renew your domain registration and could end your DNS service at any time.

If the registrar for your domain is also the DNS service provider for the domain and you don't transfer DNS service to another provider, your website, email, and the web applications associated with the domain might become unavailable.

If the transfer is successful, this method returns an operation ID that you can use to track the progress and completion of the action. If the transfer doesn't complete successfully, the domain registrant will be notified by email.

Creating a Request

data TransferDomain

The TransferDomain request includes the following elements.

See: newTransferDomain smart constructor.

Constructors

TransferDomain'

Fields

privacyProtectTechContact :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

Default: true

privacyProtectRegistrantContact :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

Default: true

autoRenew :: Maybe Bool

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Default: true

privacyProtectAdminContact :: Maybe Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

Default: true

idnLangCode :: Maybe Text

Reserved for future use.

authCode :: Maybe (Sensitive Text)

The authorization code for the domain. You get this value from the current registrar.

nameservers :: Maybe [Nameserver]

Contains details for the host and glue IP addresses.

domainName :: Text

The name of the domain that you want to transfer to Route 53. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

durationInYears :: Natural

The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain.

Default: 1

adminContact :: Sensitive ContactDetail

Provides detailed contact information.

registrantContact :: Sensitive ContactDetail

Provides detailed contact information.

techContact :: Sensitive ContactDetail

Provides detailed contact information.

newTransferDomain

Arguments

:: Text
$sel:domainName:TransferDomain'

-> Natural
$sel:durationInYears:TransferDomain'

-> ContactDetail
$sel:adminContact:TransferDomain'

-> ContactDetail
$sel:registrantContact:TransferDomain'

-> ContactDetail
$sel:techContact:TransferDomain'

-> TransferDomain

Create a value of TransferDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:privacyProtectTechContact:TransferDomain', transferDomain_privacyProtectTechContact - Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

Default: true

$sel:privacyProtectRegistrantContact:TransferDomain', transferDomain_privacyProtectRegistrantContact - Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

Default: true

$sel:autoRenew:TransferDomain', transferDomain_autoRenew - Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Default: true

$sel:privacyProtectAdminContact:TransferDomain', transferDomain_privacyProtectAdminContact - Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

Default: true

$sel:idnLangCode:TransferDomain', transferDomain_idnLangCode - Reserved for future use.

$sel:authCode:TransferDomain', transferDomain_authCode - The authorization code for the domain. You get this value from the current registrar.

$sel:nameservers:TransferDomain', transferDomain_nameservers - Contains details for the host and glue IP addresses.

$sel:domainName:TransferDomain', transferDomain_domainName - The name of the domain that you want to transfer to Route 53. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

$sel:durationInYears:TransferDomain', transferDomain_durationInYears - The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain.

Default: 1

$sel:adminContact:TransferDomain', transferDomain_adminContact - Provides detailed contact information.

$sel:registrantContact:TransferDomain', transferDomain_registrantContact - Provides detailed contact information.

$sel:techContact:TransferDomain', transferDomain_techContact - Provides detailed contact information.

Request Lenses

transferDomain_privacyProtectTechContact :: Lens' TransferDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the technical contact.

Default: true

transferDomain_privacyProtectRegistrantContact :: Lens' TransferDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the registrant contact (domain owner).

Default: true

transferDomain_autoRenew :: Lens' TransferDomain (Maybe Bool)

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Default: true

transferDomain_privacyProtectAdminContact :: Lens' TransferDomain (Maybe Bool)

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries return contact information either for Amazon Registrar (for .com, .net, and .org domains) or for our registrar associate, Gandi (for all other TLDs). If you specify false, WHOIS queries return the information that you entered for the admin contact.

Default: true

transferDomain_idnLangCode :: Lens' TransferDomain (Maybe Text)

Reserved for future use.

transferDomain_authCode :: Lens' TransferDomain (Maybe Text)

The authorization code for the domain. You get this value from the current registrar.

transferDomain_nameservers :: Lens' TransferDomain (Maybe [Nameserver])

Contains details for the host and glue IP addresses.

transferDomain_domainName :: Lens' TransferDomain Text

The name of the domain that you want to transfer to Route 53. The top-level domain (TLD), such as .com, must be a TLD that Route 53 supports. For a list of supported TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

The domain name can contain only the following characters:

Letters a through z. Domain names are not case sensitive.

Numbers 0 through 9.

Hyphen (-). You can't specify a hyphen at the beginning or end of a label.

Period (.) to separate the labels in the name, such as the . in example.com.

transferDomain_durationInYears :: Lens' TransferDomain Natural

The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain.

Default: 1

transferDomain_adminContact :: Lens' TransferDomain ContactDetail

Provides detailed contact information.

transferDomain_registrantContact :: Lens' TransferDomain ContactDetail

Provides detailed contact information.

transferDomain_techContact :: Lens' TransferDomain ContactDetail

Provides detailed contact information.

Destructuring the Response

data TransferDomainResponse

The TransferDomain response includes the following element.

See: newTransferDomainResponse smart constructor.

Constructors

TransferDomainResponse'

Fields

httpStatus :: Int

The response's http status code.

operationId :: Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

newTransferDomainResponse

Arguments

:: Int
$sel:httpStatus:TransferDomainResponse'

-> Text
$sel:operationId:TransferDomainResponse'

-> TransferDomainResponse

Create a value of TransferDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TransferDomainResponse', transferDomainResponse_httpStatus - The response's http status code.

$sel:operationId:TransferDomainResponse', transferDomainResponse_operationId - Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

Response Lenses

transferDomainResponse_httpStatus :: Lens' TransferDomainResponse Int

The response's http status code.

transferDomainResponse_operationId :: Lens' TransferDomainResponse Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.