Amazonka.Route53Domains.TransferDomainToAnotherAwsAccount

Description

Transfers a domain from the current AWS account to another AWS account. Note the following:

The AWS account that you're transferring the domain to must accept the transfer. If the other account doesn't accept the transfer within 3 days, we cancel the transfer. See AcceptDomainTransferFromAnotherAwsAccount.

You can cancel the transfer before the other account accepts it. See CancelDomainTransferToAnotherAwsAccount.

The other account can reject the transfer. See RejectDomainTransferFromAnotherAwsAccount.

When you transfer a domain from one AWS account to another, Route 53 doesn't transfer the hosted zone that is associated with the domain. DNS resolution isn't affected if the domain and the hosted zone are owned by separate accounts, so transferring the hosted zone is optional. For information about transferring the hosted zone to another AWS account, see Migrating a Hosted Zone to a Different AWS Account in the Amazon Route 53 Developer Guide.

Use either ListOperations or GetOperationDetail to determine whether the operation succeeded. GetOperationDetail provides additional information, for example, Domain Transfer from Aws Account 111122223333 has been cancelled.

Creating a Request

data TransferDomainToAnotherAwsAccount

The TransferDomainToAnotherAwsAccount request includes the following elements.

See: newTransferDomainToAnotherAwsAccount smart constructor.

Constructors

TransferDomainToAnotherAwsAccount'

Fields

domainName :: Text

The name of the domain that you want to transfer from the current AWS account to another account.

accountId :: Text

The account ID of the AWS account that you want to transfer the domain to, for example, 111122223333.

newTransferDomainToAnotherAwsAccount

Arguments

:: Text
$sel:domainName:TransferDomainToAnotherAwsAccount'

-> Text
$sel:accountId:TransferDomainToAnotherAwsAccount'

-> TransferDomainToAnotherAwsAccount

Create a value of TransferDomainToAnotherAwsAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:TransferDomainToAnotherAwsAccount', transferDomainToAnotherAwsAccount_domainName - The name of the domain that you want to transfer from the current AWS account to another account.

$sel:accountId:TransferDomainToAnotherAwsAccount', transferDomainToAnotherAwsAccount_accountId - The account ID of the AWS account that you want to transfer the domain to, for example, 111122223333.

Request Lenses

transferDomainToAnotherAwsAccount_domainName :: Lens' TransferDomainToAnotherAwsAccount Text

The name of the domain that you want to transfer from the current AWS account to another account.

transferDomainToAnotherAwsAccount_accountId :: Lens' TransferDomainToAnotherAwsAccount Text

The account ID of the AWS account that you want to transfer the domain to, for example, 111122223333.

Destructuring the Response

data TransferDomainToAnotherAwsAccountResponse

The TransferDomainToAnotherAwsAccount response includes the following elements.

See: newTransferDomainToAnotherAwsAccountResponse smart constructor.

Constructors

TransferDomainToAnotherAwsAccountResponse'

Fields

password :: Maybe Text

To finish transferring a domain to another AWS account, the account that the domain is being transferred to must submit an AcceptDomainTransferFromAnotherAwsAccount request. The request must include the value of the Password element that was returned in the TransferDomainToAnotherAwsAccount response.

operationId :: Maybe Text

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

httpStatus :: Int

The response's http status code.

newTransferDomainToAnotherAwsAccountResponse

Arguments

:: Int
$sel:httpStatus:TransferDomainToAnotherAwsAccountResponse'

-> TransferDomainToAnotherAwsAccountResponse

Create a value of TransferDomainToAnotherAwsAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:password:TransferDomainToAnotherAwsAccountResponse', transferDomainToAnotherAwsAccountResponse_password - To finish transferring a domain to another AWS account, the account that the domain is being transferred to must submit an AcceptDomainTransferFromAnotherAwsAccount request. The request must include the value of the Password element that was returned in the TransferDomainToAnotherAwsAccount response.

$sel:operationId:TransferDomainToAnotherAwsAccountResponse', transferDomainToAnotherAwsAccountResponse_operationId - Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

$sel:httpStatus:TransferDomainToAnotherAwsAccountResponse', transferDomainToAnotherAwsAccountResponse_httpStatus - The response's http status code.

Response Lenses

transferDomainToAnotherAwsAccountResponse_password :: Lens' TransferDomainToAnotherAwsAccountResponse (Maybe Text)

To finish transferring a domain to another AWS account, the account that the domain is being transferred to must submit an AcceptDomainTransferFromAnotherAwsAccount request. The request must include the value of the Password element that was returned in the TransferDomainToAnotherAwsAccount response.

transferDomainToAnotherAwsAccountResponse_operationId :: Lens' TransferDomainToAnotherAwsAccountResponse (Maybe Text)

Identifier for tracking the progress of the request. To query the operation status, use GetOperationDetail.

transferDomainToAnotherAwsAccountResponse_httpStatus :: Lens' TransferDomainToAnotherAwsAccountResponse Int

The response's http status code.