Amazonka.Route53Domains.Types.ContactDetail

Description

Documentation

data ContactDetail

ContactDetail includes the following elements.

See: newContactDetail smart constructor.

Constructors

ContactDetail'

Fields

organizationName :: Maybe Text

Name of the organization for contact types other than PERSON.

email :: Maybe Text

Email address of the contact.

state :: Maybe Text

The state or province of the contact's city.

fax :: Maybe Text

Fax number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890".

lastName :: Maybe Text

Last name of contact.

extraParams :: Maybe [ExtraParam]

A list of name-value pairs for parameters required by certain top-level domains.

zipCode :: Maybe Text

The zip or postal code of the contact's address.

addressLine1 :: Maybe Text

First line of the contact's address.

city :: Maybe Text

The city of the contact's address.

phoneNumber :: Maybe Text

The phone number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code>]". For example, a US phone number might appear as "+1.1234567890".

addressLine2 :: Maybe Text

Second line of contact's address, if any.

firstName :: Maybe Text

First name of contact.

countryCode :: Maybe CountryCode

Code for the country of the contact's address.

contactType :: Maybe ContactType

Indicates whether the contact is a person, company, association, or public organization. Note the following:

If you specify a value other than PERSON, you must also specify a value for OrganizationName.

For some TLDs, the privacy protection available depends on the value that you specify for Contact Type. For the privacy protection settings for your TLD, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide

For .es domains, if you specify PERSON, you must specify INDIVIDUAL for the value of ES_LEGAL_FORM.

newContactDetail :: ContactDetail

Create a value of ContactDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationName:ContactDetail', contactDetail_organizationName - Name of the organization for contact types other than PERSON.

$sel:email:ContactDetail', contactDetail_email - Email address of the contact.

$sel:state:ContactDetail', contactDetail_state - The state or province of the contact's city.

$sel:fax:ContactDetail', contactDetail_fax - Fax number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890".

$sel:lastName:ContactDetail', contactDetail_lastName - Last name of contact.

$sel:extraParams:ContactDetail', contactDetail_extraParams - A list of name-value pairs for parameters required by certain top-level domains.

$sel:zipCode:ContactDetail', contactDetail_zipCode - The zip or postal code of the contact's address.

$sel:addressLine1:ContactDetail', contactDetail_addressLine1 - First line of the contact's address.

$sel:city:ContactDetail', contactDetail_city - The city of the contact's address.

$sel:phoneNumber:ContactDetail', contactDetail_phoneNumber - The phone number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code>]". For example, a US phone number might appear as "+1.1234567890".

$sel:addressLine2:ContactDetail', contactDetail_addressLine2 - Second line of contact's address, if any.

$sel:firstName:ContactDetail', contactDetail_firstName - First name of contact.

$sel:countryCode:ContactDetail', contactDetail_countryCode - Code for the country of the contact's address.

$sel:contactType:ContactDetail', contactDetail_contactType - Indicates whether the contact is a person, company, association, or public organization. Note the following:

If you specify a value other than PERSON, you must also specify a value for OrganizationName.

For some TLDs, the privacy protection available depends on the value that you specify for Contact Type. For the privacy protection settings for your TLD, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide

For .es domains, if you specify PERSON, you must specify INDIVIDUAL for the value of ES_LEGAL_FORM.

contactDetail_organizationName :: Lens' ContactDetail (Maybe Text)

Name of the organization for contact types other than PERSON.

contactDetail_email :: Lens' ContactDetail (Maybe Text)

Email address of the contact.

contactDetail_state :: Lens' ContactDetail (Maybe Text)

The state or province of the contact's city.

contactDetail_fax :: Lens' ContactDetail (Maybe Text)

Fax number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code]". For example, a US phone number might appear as "+1.1234567890".

contactDetail_lastName :: Lens' ContactDetail (Maybe Text)

Last name of contact.

contactDetail_extraParams :: Lens' ContactDetail (Maybe [ExtraParam])

A list of name-value pairs for parameters required by certain top-level domains.

contactDetail_zipCode :: Lens' ContactDetail (Maybe Text)

The zip or postal code of the contact's address.

contactDetail_addressLine1 :: Lens' ContactDetail (Maybe Text)

First line of the contact's address.

contactDetail_city :: Lens' ContactDetail (Maybe Text)

The city of the contact's address.

contactDetail_phoneNumber :: Lens' ContactDetail (Maybe Text)

The phone number of the contact.

Constraints: Phone number must be specified in the format "+[country dialing code].[number including any area code>]". For example, a US phone number might appear as "+1.1234567890".

contactDetail_addressLine2 :: Lens' ContactDetail (Maybe Text)

Second line of contact's address, if any.

contactDetail_firstName :: Lens' ContactDetail (Maybe Text)

First name of contact.

contactDetail_countryCode :: Lens' ContactDetail (Maybe CountryCode)

Code for the country of the contact's address.

contactDetail_contactType :: Lens' ContactDetail (Maybe ContactType)

Indicates whether the contact is a person, company, association, or public organization. Note the following:

If you specify a value other than PERSON, you must also specify a value for OrganizationName.

For some TLDs, the privacy protection available depends on the value that you specify for Contact Type. For the privacy protection settings for your TLD, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide

For .es domains, if you specify PERSON, you must specify INDIVIDUAL for the value of ES_LEGAL_FORM.