Amazonka.Route53RecoveryCluster

Description

Derived from API version 2019-12-02 of the AWS service descriptions, licensed under Apache 2.0.

Welcome to the Amazon Route 53 Application Recovery Controller API Reference Guide for Recovery Control Data Plane .

Recovery control in Route 53 Application Recovery Controller includes extremely reliable routing controls that enable you to recover applications by rerouting traffic, for example, across Availability Zones or AWS Regions. Routing controls are simple on/off switches hosted on a cluster. A cluster is a set of five redundant regional endpoints against which you can execute API calls to update or get the state of routing controls. You use routing controls to failover traffic to recover your application across Availability Zones or Regions.

This API guide includes information about how to get and update routing control states in Route 53 Application Recovery Controller.

For more information about Route 53 Application Recovery Controller, see the following:

You can create clusters, routing controls, and control panels by using the control plane API for Recovery Control. For more information, see Amazon Route 53 Application Recovery Controller Recovery Control API Reference.

Route 53 Application Recovery Controller also provides continuous readiness checks to ensure that your applications are scaled to handle failover traffic. For more information about the related API actions, see Amazon Route 53 Application Recovery Controller Recovery Readiness API Reference.

For more information about creating resilient applications and preparing for recovery readiness with Route 53 Application Recovery Controller, see the Amazon Route 53 Application Recovery Controller Developer Guide.

Service Configuration

defaultService :: Service

API version 2019-12-02 of the Amazon Route53 Recovery Cluster SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Route53RecoveryCluster.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

There was a validation error on the request.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You don't have sufficient permissions to query the routing control state.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

There was a conflict with this request. Try again.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied because of request throttling.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

There was an unexpected error during processing of the request.

EndpointTemporarilyUnavailableException

_EndpointTemporarilyUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The cluster endpoint isn't available. Try another cluster endpoint.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request references a routing control that was not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

UpdateRoutingControlState

data UpdateRoutingControlState

See: newUpdateRoutingControlState smart constructor.

Constructors

UpdateRoutingControlState' Text RoutingControlState

newUpdateRoutingControlState

Arguments

:: Text
$sel:routingControlArn:UpdateRoutingControlState'

-> RoutingControlState
$sel:routingControlState:UpdateRoutingControlState'

-> UpdateRoutingControlState

Create a value of UpdateRoutingControlState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingControlArn:UpdateRoutingControlState', updateRoutingControlState_routingControlArn - The Amazon Resource Number (ARN) for the routing control that you want to update the state for.

$sel:routingControlState:UpdateRoutingControlState', updateRoutingControlState_routingControlState - The state of the routing control. You can set the value to be On or Off.

data UpdateRoutingControlStateResponse

See: newUpdateRoutingControlStateResponse smart constructor.

Constructors

UpdateRoutingControlStateResponse' Int

newUpdateRoutingControlStateResponse

Arguments

:: Int
$sel:httpStatus:UpdateRoutingControlStateResponse'

-> UpdateRoutingControlStateResponse

Create a value of UpdateRoutingControlStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateRoutingControlStateResponse', updateRoutingControlStateResponse_httpStatus - The response's http status code.

GetRoutingControlState

data GetRoutingControlState

See: newGetRoutingControlState smart constructor.

Constructors

GetRoutingControlState' Text

newGetRoutingControlState

Arguments

:: Text
$sel:routingControlArn:GetRoutingControlState'

-> GetRoutingControlState

Create a value of GetRoutingControlState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingControlArn:GetRoutingControlState', getRoutingControlState_routingControlArn - The Amazon Resource Number (ARN) for the routing control that you want to get the state for.

data GetRoutingControlStateResponse

See: newGetRoutingControlStateResponse smart constructor.

Constructors

GetRoutingControlStateResponse' Int Text RoutingControlState

newGetRoutingControlStateResponse

Arguments

:: Int
$sel:httpStatus:GetRoutingControlStateResponse'

-> Text
$sel:routingControlArn:GetRoutingControlState'

-> RoutingControlState
$sel:routingControlState:GetRoutingControlStateResponse'

-> GetRoutingControlStateResponse

Create a value of GetRoutingControlStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetRoutingControlStateResponse', getRoutingControlStateResponse_httpStatus - The response's http status code.

$sel:routingControlArn:GetRoutingControlState', getRoutingControlStateResponse_routingControlArn - The Amazon Resource Number (ARN) of the response.

$sel:routingControlState:GetRoutingControlStateResponse', getRoutingControlStateResponse_routingControlState - The state of the routing control.

UpdateRoutingControlStates

data UpdateRoutingControlStates

See: newUpdateRoutingControlStates smart constructor.

Constructors

UpdateRoutingControlStates' [UpdateRoutingControlStateEntry]

newUpdateRoutingControlStates :: UpdateRoutingControlStates

Create a value of UpdateRoutingControlStates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateRoutingControlStateEntries:UpdateRoutingControlStates', updateRoutingControlStates_updateRoutingControlStateEntries - A set of routing control entries that you want to update.

data UpdateRoutingControlStatesResponse

See: newUpdateRoutingControlStatesResponse smart constructor.

Constructors

UpdateRoutingControlStatesResponse' Int

newUpdateRoutingControlStatesResponse

Arguments

:: Int
$sel:httpStatus:UpdateRoutingControlStatesResponse'

-> UpdateRoutingControlStatesResponse

Create a value of UpdateRoutingControlStatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateRoutingControlStatesResponse', updateRoutingControlStatesResponse_httpStatus - The response's http status code.

Types

RoutingControlState

newtype RoutingControlState

Constructors

RoutingControlState'

Fields

fromRoutingControlState :: Text

Bundled Patterns

pattern RoutingControlState_Off :: RoutingControlState

pattern RoutingControlState_On :: RoutingControlState

UpdateRoutingControlStateEntry

data UpdateRoutingControlStateEntry

A routing control state.

See: newUpdateRoutingControlStateEntry smart constructor.

Constructors

UpdateRoutingControlStateEntry' Text RoutingControlState

newUpdateRoutingControlStateEntry

Arguments

:: Text
$sel:routingControlArn:UpdateRoutingControlStateEntry'

-> RoutingControlState
$sel:routingControlState:UpdateRoutingControlStateEntry'

-> UpdateRoutingControlStateEntry

Create a value of UpdateRoutingControlStateEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingControlArn:UpdateRoutingControlStateEntry', updateRoutingControlStateEntry_routingControlArn - The Amazon Resource Number (ARN) for the routing control state entry.

$sel:routingControlState:UpdateRoutingControlStateEntry', updateRoutingControlStateEntry_routingControlState - The routing control state in a set of routing control state entries.