Amazonka.Route53RecoveryCluster.Lens

Description

Operations

UpdateRoutingControlState

updateRoutingControlState_routingControlArn :: Lens' UpdateRoutingControlState Text

The Amazon Resource Number (ARN) for the routing control that you want to update the state for.

updateRoutingControlState_routingControlState :: Lens' UpdateRoutingControlState RoutingControlState

The state of the routing control. You can set the value to be On or Off.

updateRoutingControlStateResponse_httpStatus :: Lens' UpdateRoutingControlStateResponse Int

The response's http status code.

GetRoutingControlState

getRoutingControlState_routingControlArn :: Lens' GetRoutingControlState Text

The Amazon Resource Number (ARN) for the routing control that you want to get the state for.

getRoutingControlStateResponse_httpStatus :: Lens' GetRoutingControlStateResponse Int

The response's http status code.

getRoutingControlStateResponse_routingControlArn :: Lens' GetRoutingControlStateResponse Text

The Amazon Resource Number (ARN) of the response.

getRoutingControlStateResponse_routingControlState :: Lens' GetRoutingControlStateResponse RoutingControlState

The state of the routing control.

UpdateRoutingControlStates

updateRoutingControlStates_updateRoutingControlStateEntries :: Lens' UpdateRoutingControlStates [UpdateRoutingControlStateEntry]

A set of routing control entries that you want to update.

updateRoutingControlStatesResponse_httpStatus :: Lens' UpdateRoutingControlStatesResponse Int

The response's http status code.

Types

UpdateRoutingControlStateEntry

updateRoutingControlStateEntry_routingControlArn :: Lens' UpdateRoutingControlStateEntry Text

The Amazon Resource Number (ARN) for the routing control state entry.

updateRoutingControlStateEntry_routingControlState :: Lens' UpdateRoutingControlStateEntry RoutingControlState

The routing control state in a set of routing control state entries.