Amazonka.Route53RecoveryControlConfig.Lens

Description

Operations

DescribeControlPanel

describeControlPanel_controlPanelArn :: Lens' DescribeControlPanel Text

The Amazon Resource Name (ARN) of the control panel that you're getting details for.

describeControlPanelResponse_controlPanel :: Lens' DescribeControlPanelResponse (Maybe ControlPanel)

Information about the control panel.

describeControlPanelResponse_httpStatus :: Lens' DescribeControlPanelResponse Int

The response's http status code.

CreateRoutingControl

createRoutingControl_controlPanelArn :: Lens' CreateRoutingControl (Maybe Text)

The Amazon Resource Name (ARN) of the control panel that includes the routing control.

createRoutingControl_clientToken :: Lens' CreateRoutingControl (Maybe Text)

Unique client idempotency token.

createRoutingControl_clusterArn :: Lens' CreateRoutingControl Text

The Amazon Resource Name (ARN) of the cluster that includes the routing control.

createRoutingControl_routingControlName :: Lens' CreateRoutingControl Text

The name of the routing control.

createRoutingControlResponse_routingControl :: Lens' CreateRoutingControlResponse (Maybe RoutingControl)

The routing control that is created.

createRoutingControlResponse_httpStatus :: Lens' CreateRoutingControlResponse Int

The response's http status code.

DescribeCluster

describeCluster_clusterArn :: Lens' DescribeCluster Text

The Amazon Resource Name (ARN) of the cluster that you're getting details for.

describeClusterResponse_cluster :: Lens' DescribeClusterResponse (Maybe Cluster)

The cluster for the DescribeCluster request.

describeClusterResponse_httpStatus :: Lens' DescribeClusterResponse Int

The response's http status code.

DeleteRoutingControl

deleteRoutingControl_routingControlArn :: Lens' DeleteRoutingControl Text

The Amazon Resource Name (ARN) of the routing control that you're deleting.

deleteRoutingControlResponse_httpStatus :: Lens' DeleteRoutingControlResponse Int

The response's http status code.

UpdateRoutingControl

updateRoutingControl_routingControlName :: Lens' UpdateRoutingControl Text

The name of the routing control.

updateRoutingControl_routingControlArn :: Lens' UpdateRoutingControl Text

The Amazon Resource Name (ARN) of the routing control.

updateRoutingControlResponse_routingControl :: Lens' UpdateRoutingControlResponse (Maybe RoutingControl)

The routing control that was updated.

updateRoutingControlResponse_httpStatus :: Lens' UpdateRoutingControlResponse Int

The response's http status code.

CreateControlPanel

createControlPanel_clientToken :: Lens' CreateControlPanel (Maybe Text)

Unique client idempotency token.

createControlPanel_clusterArn :: Lens' CreateControlPanel Text

The Amazon Resource Name (ARN) of the cluster for the control panel.

createControlPanel_controlPanelName :: Lens' CreateControlPanel Text

The name of the control panel.

createControlPanelResponse_controlPanel :: Lens' CreateControlPanelResponse (Maybe ControlPanel)

Information about a control panel.

createControlPanelResponse_httpStatus :: Lens' CreateControlPanelResponse Int

The response's http status code.

UpdateControlPanel

updateControlPanel_controlPanelArn :: Lens' UpdateControlPanel Text

The Amazon Resource Name (ARN) of the control panel.

updateControlPanel_controlPanelName :: Lens' UpdateControlPanel Text

The name of the control panel.

updateControlPanelResponse_controlPanel :: Lens' UpdateControlPanelResponse (Maybe ControlPanel)

The control panel to update.

updateControlPanelResponse_httpStatus :: Lens' UpdateControlPanelResponse Int

The response's http status code.

DeleteControlPanel

deleteControlPanel_controlPanelArn :: Lens' DeleteControlPanel Text

The Amazon Resource Name (ARN) of the control panel that you're deleting.

deleteControlPanelResponse_httpStatus :: Lens' DeleteControlPanelResponse Int

The response's http status code.

DeleteCluster

deleteCluster_clusterArn :: Lens' DeleteCluster Text

The Amazon Resource Name (ARN) of the cluster that you're deleting.

deleteClusterResponse_httpStatus :: Lens' DeleteClusterResponse Int

The response's http status code.

CreateSafetyRule

createSafetyRule_assertionRule :: Lens' CreateSafetyRule (Maybe NewAssertionRule)

Undocumented member.

createSafetyRule_clientToken :: Lens' CreateSafetyRule (Maybe Text)

Unique client idempotency token.

createSafetyRule_gatingRule :: Lens' CreateSafetyRule (Maybe NewGatingRule)

Undocumented member.

createSafetyRuleResponse_assertionRule :: Lens' CreateSafetyRuleResponse (Maybe AssertionRule)

Undocumented member.

createSafetyRuleResponse_gatingRule :: Lens' CreateSafetyRuleResponse (Maybe GatingRule)

Undocumented member.

createSafetyRuleResponse_httpStatus :: Lens' CreateSafetyRuleResponse Int

The response's http status code.

CreateCluster

createCluster_clientToken :: Lens' CreateCluster (Maybe Text)

Unique client idempotency token.

createCluster_clusterName :: Lens' CreateCluster Text

The name of the cluster.

createClusterResponse_cluster :: Lens' CreateClusterResponse (Maybe Cluster)

The cluster that was created.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.

ListAssociatedRoute53HealthChecks

listAssociatedRoute53HealthChecks_nextToken :: Lens' ListAssociatedRoute53HealthChecks (Maybe Text)

The token that identifies which batch of results you want to see.

listAssociatedRoute53HealthChecks_maxResults :: Lens' ListAssociatedRoute53HealthChecks (Maybe Natural)

The number of objects that you want to return with this call.

listAssociatedRoute53HealthChecks_routingControlArn :: Lens' ListAssociatedRoute53HealthChecks Text

The Amazon Resource Name (ARN) of the routing control that you're getting details for.

listAssociatedRoute53HealthChecksResponse_nextToken :: Lens' ListAssociatedRoute53HealthChecksResponse (Maybe Text)

The token that identifies which batch of results you want to see.

listAssociatedRoute53HealthChecksResponse_healthCheckIds :: Lens' ListAssociatedRoute53HealthChecksResponse (Maybe [Text])

Identifiers for the health checks.

listAssociatedRoute53HealthChecksResponse_httpStatus :: Lens' ListAssociatedRoute53HealthChecksResponse Int

The response's http status code.

DescribeSafetyRule

describeSafetyRule_safetyRuleArn :: Lens' DescribeSafetyRule Text

The request body that you include when you update a safety rule.

describeSafetyRuleResponse_assertionRule :: Lens' DescribeSafetyRuleResponse (Maybe AssertionRule)

Undocumented member.

describeSafetyRuleResponse_gatingRule :: Lens' DescribeSafetyRuleResponse (Maybe GatingRule)

Undocumented member.

describeSafetyRuleResponse_httpStatus :: Lens' DescribeSafetyRuleResponse Int

The response's http status code.

ListRoutingControls

listRoutingControls_nextToken :: Lens' ListRoutingControls (Maybe Text)

The token that identifies which batch of results you want to see.

listRoutingControls_maxResults :: Lens' ListRoutingControls (Maybe Natural)

The number of objects that you want to return with this call.

listRoutingControls_controlPanelArn :: Lens' ListRoutingControls Text

The Amazon Resource Name (ARN) of the control panel that you're getting routing control details for.

listRoutingControlsResponse_nextToken :: Lens' ListRoutingControlsResponse (Maybe Text)

The token that identifies which batch of results you want to see.

listRoutingControlsResponse_routingControls :: Lens' ListRoutingControlsResponse (Maybe [RoutingControl])

An array of routing controls.

listRoutingControlsResponse_httpStatus :: Lens' ListRoutingControlsResponse Int

The response's http status code.

ListControlPanels

listControlPanels_clusterArn :: Lens' ListControlPanels (Maybe Text)

The Amazon Resource Name (ARN) of a cluster.

listControlPanels_nextToken :: Lens' ListControlPanels (Maybe Text)

The token that identifies which batch of results you want to see.

listControlPanels_maxResults :: Lens' ListControlPanels (Maybe Natural)

The number of objects that you want to return with this call.

listControlPanelsResponse_nextToken :: Lens' ListControlPanelsResponse (Maybe Text)

The token that identifies which batch of results you want to see.

listControlPanelsResponse_controlPanels :: Lens' ListControlPanelsResponse (Maybe [ControlPanel])

The result of a successful ListControlPanel request.

listControlPanelsResponse_httpStatus :: Lens' ListControlPanelsResponse Int

The response's http status code.

UpdateSafetyRule

updateSafetyRule_gatingRuleUpdate :: Lens' UpdateSafetyRule (Maybe GatingRuleUpdate)

Undocumented member.

updateSafetyRule_assertionRuleUpdate :: Lens' UpdateSafetyRule (Maybe AssertionRuleUpdate)

Undocumented member.

updateSafetyRuleResponse_assertionRule :: Lens' UpdateSafetyRuleResponse (Maybe AssertionRule)

Undocumented member.

updateSafetyRuleResponse_gatingRule :: Lens' UpdateSafetyRuleResponse (Maybe GatingRule)

Undocumented member.

updateSafetyRuleResponse_httpStatus :: Lens' UpdateSafetyRuleResponse Int

The response's http status code.

DeleteSafetyRule

deleteSafetyRule_safetyRuleArn :: Lens' DeleteSafetyRule Text

The request body that you include when you update a safety rule.

deleteSafetyRuleResponse_httpStatus :: Lens' DeleteSafetyRuleResponse Int

The response's http status code.

ListClusters

listClusters_nextToken :: Lens' ListClusters (Maybe Text)

The token that identifies which batch of results you want to see.

listClusters_maxResults :: Lens' ListClusters (Maybe Natural)

The number of objects that you want to return with this call.

listClustersResponse_nextToken :: Lens' ListClustersResponse (Maybe Text)

The token that identifies which batch of results you want to see.

listClustersResponse_clusters :: Lens' ListClustersResponse (Maybe [Cluster])

An array of the clusters in an account.

listClustersResponse_httpStatus :: Lens' ListClustersResponse Int

The response's http status code.

ListSafetyRules

listSafetyRules_nextToken :: Lens' ListSafetyRules (Maybe Text)

The token that identifies which batch of results you want to see.

listSafetyRules_maxResults :: Lens' ListSafetyRules (Maybe Natural)

The number of objects that you want to return with this call.

listSafetyRules_controlPanelArn :: Lens' ListSafetyRules Text

The Amazon Resource Name (ARN) of the control panel that you're getting details for.

listSafetyRulesResponse_nextToken :: Lens' ListSafetyRulesResponse (Maybe Text)

The token that identifies which batch of results you want to see.

listSafetyRulesResponse_safetyRules :: Lens' ListSafetyRulesResponse (Maybe [Rule])

The list of safety rules in a control panel.

listSafetyRulesResponse_httpStatus :: Lens' ListSafetyRulesResponse Int

The response's http status code.

DescribeRoutingControl

describeRoutingControl_routingControlArn :: Lens' DescribeRoutingControl Text

The Amazon Resource Name (ARN) of the routing control that you're getting details for.

describeRoutingControlResponse_routingControl :: Lens' DescribeRoutingControlResponse (Maybe RoutingControl)

Information about the routing control.

describeRoutingControlResponse_httpStatus :: Lens' DescribeRoutingControlResponse Int

The response's http status code.

Types

AssertionRule

assertionRule_status :: Lens' AssertionRule Status

The deployment status of an assertion rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

assertionRule_controlPanelArn :: Lens' AssertionRule Text

The Amazon Resource Name (ARN) of the control panel.

assertionRule_safetyRuleArn :: Lens' AssertionRule Text

The Amazon Resource Name (ARN) of the assertion rule.

assertionRule_assertedControls :: Lens' AssertionRule [Text]

The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

assertionRule_ruleConfig :: Lens' AssertionRule RuleConfig

The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

assertionRule_waitPeriodMs :: Lens' AssertionRule Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

assertionRule_name :: Lens' AssertionRule Text

Name of the assertion rule. You can use any non-white space character in the name.

AssertionRuleUpdate

assertionRuleUpdate_safetyRuleArn :: Lens' AssertionRuleUpdate Text

The Amazon Resource Name (ARN) of the assertion rule.

assertionRuleUpdate_waitPeriodMs :: Lens' AssertionRuleUpdate Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

assertionRuleUpdate_name :: Lens' AssertionRuleUpdate Text

The name of the assertion rule. You can use any non-white space character in the name.

Cluster

cluster_status :: Lens' Cluster (Maybe Status)

Deployment status of a resource. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

cluster_clusterArn :: Lens' Cluster (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

cluster_name :: Lens' Cluster (Maybe Text)

The name of the cluster.

cluster_clusterEndpoints :: Lens' Cluster (Maybe [ClusterEndpoint])

Endpoints for a cluster. Specify one of these endpoints when you want to set or retrieve a routing control state in the cluster.

To get or update the routing control state, see the Amazon Route 53 Application Recovery Controller Cluster (Data Plane) Actions.

ClusterEndpoint

clusterEndpoint_region :: Lens' ClusterEndpoint (Maybe Text)

The Amazon Web Services Region for a cluster endpoint.

clusterEndpoint_endpoint :: Lens' ClusterEndpoint (Maybe Text)

A cluster endpoint. Specify an endpoint and Amazon Web Services Region when you want to set or retrieve a routing control state in the cluster.

To get or update the routing control state, see the Amazon Route 53 Application Recovery Controller Cluster (Data Plane) Actions.

ControlPanel

controlPanel_status :: Lens' ControlPanel (Maybe Status)

The deployment status of control panel. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

controlPanel_controlPanelArn :: Lens' ControlPanel (Maybe Text)

The Amazon Resource Name (ARN) of the control panel.

controlPanel_clusterArn :: Lens' ControlPanel (Maybe Text)

The Amazon Resource Name (ARN) of the cluster that includes the control panel.

controlPanel_routingControlCount :: Lens' ControlPanel (Maybe Int)

The number of routing controls in the control panel.

controlPanel_name :: Lens' ControlPanel (Maybe Text)

The name of the control panel. You can use any non-white space character in the name.

controlPanel_defaultControlPanel :: Lens' ControlPanel (Maybe Bool)

A flag that Amazon Route 53 Application Recovery Controller sets to true to designate the default control panel for a cluster. When you create a cluster, Amazon Route 53 Application Recovery Controller creates a control panel, and sets this flag for that control panel. If you create a control panel yourself, this flag is set to false.

GatingRule

gatingRule_status :: Lens' GatingRule Status

The deployment status of a gating rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

gatingRule_targetControls :: Lens' GatingRule [Text]

Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify ATLEAST 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

gatingRule_controlPanelArn :: Lens' GatingRule Text

The Amazon Resource Name (ARN) of the control panel.

gatingRule_safetyRuleArn :: Lens' GatingRule Text

The Amazon Resource Name (ARN) of the gating rule.

gatingRule_gatingControls :: Lens' GatingRule [Text]

The gating controls for the gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

gatingRule_ruleConfig :: Lens' GatingRule RuleConfig

The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

gatingRule_waitPeriodMs :: Lens' GatingRule Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

gatingRule_name :: Lens' GatingRule Text

The name for the gating rule.

GatingRuleUpdate

gatingRuleUpdate_safetyRuleArn :: Lens' GatingRuleUpdate Text

The Amazon Resource Name (ARN) of the gating rule.

gatingRuleUpdate_waitPeriodMs :: Lens' GatingRuleUpdate Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

gatingRuleUpdate_name :: Lens' GatingRuleUpdate Text

The name for the gating rule.

NewAssertionRule

newAssertionRule_controlPanelArn :: Lens' NewAssertionRule Text

The Amazon Resource Name (ARN) for the control panel.

newAssertionRule_assertedControls :: Lens' NewAssertionRule [Text]

The routing controls that are part of transactions that are evaluated to determine if a request to change a routing control state is allowed. For example, you might include three routing controls, one for each of three Amazon Web Services Regions.

newAssertionRule_ruleConfig :: Lens' NewAssertionRule RuleConfig

The criteria that you set for specific assertion controls (routing controls) that designate how many controls must be enabled as the result of a transaction. For example, if you have three assertion controls, you might specify atleast 2 for your rule configuration. This means that at least two assertion controls must be enabled, so that at least two Amazon Web Services Regions are enabled.

newAssertionRule_waitPeriodMs :: Lens' NewAssertionRule Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

newAssertionRule_name :: Lens' NewAssertionRule Text

The name of the assertion rule. You can use any non-white space character in the name.

NewGatingRule

newGatingRule_targetControls :: Lens' NewGatingRule [Text]

Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify AtLeast 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

newGatingRule_controlPanelArn :: Lens' NewGatingRule Text

The Amazon Resource Name (ARN) of the control panel.

newGatingRule_gatingControls :: Lens' NewGatingRule [Text]

The gating controls for the new gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

newGatingRule_ruleConfig :: Lens' NewGatingRule RuleConfig

The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

newGatingRule_waitPeriodMs :: Lens' NewGatingRule Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

newGatingRule_name :: Lens' NewGatingRule Text

The name for the new gating rule.

RoutingControl

routingControl_status :: Lens' RoutingControl (Maybe Status)

The deployment status of a routing control. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

routingControl_controlPanelArn :: Lens' RoutingControl (Maybe Text)

The Amazon Resource Name (ARN) of the control panel that includes the routing control.

routingControl_name :: Lens' RoutingControl (Maybe Text)

The name of the routing control.

routingControl_routingControlArn :: Lens' RoutingControl (Maybe Text)

The Amazon Resource Name (ARN) of the routing control.

Rule

rule_gating :: Lens' Rule (Maybe GatingRule)

A gating rule verifies that a set of gating controls evaluates as true, based on a rule configuration that you specify. If the gating rule evaluates to true, Amazon Route 53 Application Recovery Controller allows a set of routing control state changes to run and complete against the set of target controls.

rule_assertion :: Lens' Rule (Maybe AssertionRule)

An assertion rule enforces that, when a routing control state is changed, the criteria set by the rule configuration is met. Otherwise, the change to the routing control is not accepted.

RuleConfig

ruleConfig_type :: Lens' RuleConfig RuleType

A rule can be one of the following: ATLEAST, AND, or OR.

ruleConfig_inverted :: Lens' RuleConfig Bool

Logical negation of the rule. If the rule would usually evaluate true, it's evaluated as false, and vice versa.

ruleConfig_threshold :: Lens' RuleConfig Int

The value of N, when you specify an ATLEAST rule type. That is, Threshold is the number of controls that must be set when you specify an ATLEAST type.