Amazonka.Route53RecoveryControlConfig.ListControlPanels

Description

Returns an array of control panels for a cluster.

Creating a Request

data ListControlPanels

See: newListControlPanels smart constructor.

Constructors

ListControlPanels'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of a cluster.

nextToken :: Maybe Text

The token that identifies which batch of results you want to see.

maxResults :: Maybe Natural

The number of objects that you want to return with this call.

newListControlPanels :: ListControlPanels

Create a value of ListControlPanels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:ListControlPanels', listControlPanels_clusterArn - The Amazon Resource Name (ARN) of a cluster.

$sel:nextToken:ListControlPanels', listControlPanels_nextToken - The token that identifies which batch of results you want to see.

$sel:maxResults:ListControlPanels', listControlPanels_maxResults - The number of objects that you want to return with this call.

Request Lenses

listControlPanels_clusterArn :: Lens' ListControlPanels (Maybe Text)

The Amazon Resource Name (ARN) of a cluster.

listControlPanels_nextToken :: Lens' ListControlPanels (Maybe Text)

The token that identifies which batch of results you want to see.

listControlPanels_maxResults :: Lens' ListControlPanels (Maybe Natural)

The number of objects that you want to return with this call.

Destructuring the Response

data ListControlPanelsResponse

See: newListControlPanelsResponse smart constructor.

Constructors

ListControlPanelsResponse'

Fields

nextToken :: Maybe Text

The token that identifies which batch of results you want to see.

controlPanels :: Maybe [ControlPanel]

The result of a successful ListControlPanel request.

httpStatus :: Int

The response's http status code.

newListControlPanelsResponse

Arguments

:: Int
$sel:httpStatus:ListControlPanelsResponse'

-> ListControlPanelsResponse

Create a value of ListControlPanelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListControlPanels', listControlPanelsResponse_nextToken - The token that identifies which batch of results you want to see.

$sel:controlPanels:ListControlPanelsResponse', listControlPanelsResponse_controlPanels - The result of a successful ListControlPanel request.

$sel:httpStatus:ListControlPanelsResponse', listControlPanelsResponse_httpStatus - The response's http status code.

Response Lenses

listControlPanelsResponse_nextToken :: Lens' ListControlPanelsResponse (Maybe Text)

The token that identifies which batch of results you want to see.

listControlPanelsResponse_controlPanels :: Lens' ListControlPanelsResponse (Maybe [ControlPanel])

The result of a successful ListControlPanel request.

listControlPanelsResponse_httpStatus :: Lens' ListControlPanelsResponse Int

The response's http status code.