Amazonka.Route53RecoveryControlConfig.Types.ControlPanel

Description

Documentation

data ControlPanel

A control panel represents a group of routing controls that can be changed together in a single transaction.

See: newControlPanel smart constructor.

Constructors

ControlPanel'

Fields

status :: Maybe Status

The deployment status of control panel. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

controlPanelArn :: Maybe Text

The Amazon Resource Name (ARN) of the control panel.

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster that includes the control panel.

routingControlCount :: Maybe Int

The number of routing controls in the control panel.

name :: Maybe Text

The name of the control panel. You can use any non-white space character in the name.

defaultControlPanel :: Maybe Bool

A flag that Amazon Route 53 Application Recovery Controller sets to true to designate the default control panel for a cluster. When you create a cluster, Amazon Route 53 Application Recovery Controller creates a control panel, and sets this flag for that control panel. If you create a control panel yourself, this flag is set to false.

newControlPanel :: ControlPanel

Create a value of ControlPanel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ControlPanel', controlPanel_status - The deployment status of control panel. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

$sel:controlPanelArn:ControlPanel', controlPanel_controlPanelArn - The Amazon Resource Name (ARN) of the control panel.

$sel:clusterArn:ControlPanel', controlPanel_clusterArn - The Amazon Resource Name (ARN) of the cluster that includes the control panel.

$sel:routingControlCount:ControlPanel', controlPanel_routingControlCount - The number of routing controls in the control panel.

$sel:name:ControlPanel', controlPanel_name - The name of the control panel. You can use any non-white space character in the name.

$sel:defaultControlPanel:ControlPanel', controlPanel_defaultControlPanel - A flag that Amazon Route 53 Application Recovery Controller sets to true to designate the default control panel for a cluster. When you create a cluster, Amazon Route 53 Application Recovery Controller creates a control panel, and sets this flag for that control panel. If you create a control panel yourself, this flag is set to false.

controlPanel_status :: Lens' ControlPanel (Maybe Status)

The deployment status of control panel. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

controlPanel_controlPanelArn :: Lens' ControlPanel (Maybe Text)

The Amazon Resource Name (ARN) of the control panel.

controlPanel_clusterArn :: Lens' ControlPanel (Maybe Text)

The Amazon Resource Name (ARN) of the cluster that includes the control panel.

controlPanel_routingControlCount :: Lens' ControlPanel (Maybe Int)

The number of routing controls in the control panel.

controlPanel_name :: Lens' ControlPanel (Maybe Text)

The name of the control panel. You can use any non-white space character in the name.

controlPanel_defaultControlPanel :: Lens' ControlPanel (Maybe Bool)

A flag that Amazon Route 53 Application Recovery Controller sets to true to designate the default control panel for a cluster. When you create a cluster, Amazon Route 53 Application Recovery Controller creates a control panel, and sets this flag for that control panel. If you create a control panel yourself, this flag is set to false.