Amazonka.Route53RecoveryControlConfig.Types.GatingRule

Description

Documentation

data GatingRule

A gating rule verifies that a set of gating controls evaluates as true, based on a rule configuration that you specify. If the gating rule evaluates to true, Amazon Route 53 Application Recovery Controller allows a set of routing control state changes to run and complete against the set of target controls.

See: newGatingRule smart constructor.

Constructors

GatingRule'

Fields

status :: Status

The deployment status of a gating rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

targetControls :: [Text]

Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify ATLEAST 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

controlPanelArn :: Text

The Amazon Resource Name (ARN) of the control panel.

safetyRuleArn :: Text

The Amazon Resource Name (ARN) of the gating rule.

gatingControls :: [Text]

The gating controls for the gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

ruleConfig :: RuleConfig

The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

waitPeriodMs :: Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

name :: Text

The name for the gating rule.

newGatingRule

Arguments

:: Status
$sel:status:GatingRule'

-> Text
$sel:controlPanelArn:GatingRule'

-> Text
$sel:safetyRuleArn:GatingRule'

-> RuleConfig
$sel:ruleConfig:GatingRule'

-> Int
$sel:waitPeriodMs:GatingRule'

-> Text
$sel:name:GatingRule'

-> GatingRule

Create a value of GatingRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GatingRule', gatingRule_status - The deployment status of a gating rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

$sel:targetControls:GatingRule', gatingRule_targetControls - Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify ATLEAST 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

$sel:controlPanelArn:GatingRule', gatingRule_controlPanelArn - The Amazon Resource Name (ARN) of the control panel.

$sel:safetyRuleArn:GatingRule', gatingRule_safetyRuleArn - The Amazon Resource Name (ARN) of the gating rule.

$sel:gatingControls:GatingRule', gatingRule_gatingControls - The gating controls for the gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

$sel:ruleConfig:GatingRule', gatingRule_ruleConfig - The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

$sel:waitPeriodMs:GatingRule', gatingRule_waitPeriodMs - An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

$sel:name:GatingRule', gatingRule_name - The name for the gating rule.

gatingRule_status :: Lens' GatingRule Status

The deployment status of a gating rule. Status can be one of the following: PENDING, DEPLOYED, PENDING_DELETION.

gatingRule_targetControls :: Lens' GatingRule [Text]

Routing controls that can only be set or unset if the specified RuleConfig evaluates to true for the specified GatingControls. For example, say you have three gating controls, one for each of three Amazon Web Services Regions. Now you specify ATLEAST 2 as your RuleConfig. With these settings, you can only change (set or unset) the routing controls that you have specified as TargetControls if that rule evaluates to true.

In other words, your ability to change the routing controls that you have specified as TargetControls is gated by the rule that you set for the routing controls in GatingControls.

gatingRule_controlPanelArn :: Lens' GatingRule Text

The Amazon Resource Name (ARN) of the control panel.

gatingRule_safetyRuleArn :: Lens' GatingRule Text

The Amazon Resource Name (ARN) of the gating rule.

gatingRule_gatingControls :: Lens' GatingRule [Text]

The gating controls for the gating rule. That is, routing controls that are evaluated by the rule configuration that you specify.

gatingRule_ruleConfig :: Lens' GatingRule RuleConfig

The criteria that you set for specific gating controls (routing controls) that designates how many controls must be enabled to allow you to change (set or unset) the target controls.

gatingRule_waitPeriodMs :: Lens' GatingRule Int

An evaluation period, in milliseconds (ms), during which any request against the target routing controls will fail. This helps prevent "flapping" of state. The wait period is 5000 ms by default, but you can choose a custom value.

gatingRule_name :: Lens' GatingRule Text

The name for the gating rule.