Amazonka.Route53RecoveryReadiness.CreateReadinessCheck

Description

Creates a new Readiness Check.

Creating a Request

data CreateReadinessCheck

The ReadinessCheck to create

See: newCreateReadinessCheck smart constructor.

Constructors

CreateReadinessCheck'

Fields

tags :: Maybe (HashMap Text Text)

resourceSetName :: Text

The name of the ResourceSet to check

readinessCheckName :: Text

The name of the ReadinessCheck to create

newCreateReadinessCheck

Arguments

:: Text
$sel:resourceSetName:CreateReadinessCheck'

-> Text
$sel:readinessCheckName:CreateReadinessCheck'

-> CreateReadinessCheck

Create a value of CreateReadinessCheck with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateReadinessCheck', createReadinessCheck_tags - Undocumented member.

$sel:resourceSetName:CreateReadinessCheck', createReadinessCheck_resourceSetName - The name of the ResourceSet to check

$sel:readinessCheckName:CreateReadinessCheck', createReadinessCheck_readinessCheckName - The name of the ReadinessCheck to create

Request Lenses

createReadinessCheck_tags :: Lens' CreateReadinessCheck (Maybe (HashMap Text Text))

Undocumented member.

createReadinessCheck_resourceSetName :: Lens' CreateReadinessCheck Text

The name of the ResourceSet to check

createReadinessCheck_readinessCheckName :: Lens' CreateReadinessCheck Text

The name of the ReadinessCheck to create

Destructuring the Response

data CreateReadinessCheckResponse

See: newCreateReadinessCheckResponse smart constructor.

Constructors

CreateReadinessCheckResponse'

Fields

readinessCheckName :: Maybe Text

Name for a ReadinessCheck

resourceSet :: Maybe Text

Name of the ResourceSet to be checked

readinessCheckArn :: Maybe Text

Arn associated with ReadinessCheck

tags :: Maybe (HashMap Text Text)

httpStatus :: Int

The response's http status code.

newCreateReadinessCheckResponse

Arguments

:: Int
$sel:httpStatus:CreateReadinessCheckResponse'

-> CreateReadinessCheckResponse

Create a value of CreateReadinessCheckResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readinessCheckName:CreateReadinessCheck', createReadinessCheckResponse_readinessCheckName - Name for a ReadinessCheck

$sel:resourceSet:CreateReadinessCheckResponse', createReadinessCheckResponse_resourceSet - Name of the ResourceSet to be checked

$sel:readinessCheckArn:CreateReadinessCheckResponse', createReadinessCheckResponse_readinessCheckArn - Arn associated with ReadinessCheck

$sel:tags:CreateReadinessCheck', createReadinessCheckResponse_tags - Undocumented member.

$sel:httpStatus:CreateReadinessCheckResponse', createReadinessCheckResponse_httpStatus - The response's http status code.

Response Lenses

createReadinessCheckResponse_readinessCheckName :: Lens' CreateReadinessCheckResponse (Maybe Text)

Name for a ReadinessCheck

createReadinessCheckResponse_resourceSet :: Lens' CreateReadinessCheckResponse (Maybe Text)

Name of the ResourceSet to be checked

createReadinessCheckResponse_readinessCheckArn :: Lens' CreateReadinessCheckResponse (Maybe Text)

Arn associated with ReadinessCheck

createReadinessCheckResponse_tags :: Lens' CreateReadinessCheckResponse (Maybe (HashMap Text Text))

Undocumented member.

createReadinessCheckResponse_httpStatus :: Lens' CreateReadinessCheckResponse Int

The response's http status code.