Amazonka.Route53RecoveryReadiness.CreateResourceSet

Description

Creates a new Resource Set.

Creating a Request

data CreateResourceSet

The ResourceSet to create

See: newCreateResourceSet smart constructor.

Constructors

CreateResourceSet'

Fields

tags :: Maybe (HashMap Text Text)

resourceSetType :: Text

AWS Resource type of the resources in the ResourceSet

resourceSetName :: Text

The name of the ResourceSet to create

resources :: [Resource]

A list of Resource objects

newCreateResourceSet

Arguments

:: Text
$sel:resourceSetType:CreateResourceSet'

-> Text
$sel:resourceSetName:CreateResourceSet'

-> CreateResourceSet

Create a value of CreateResourceSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateResourceSet', createResourceSet_tags - Undocumented member.

$sel:resourceSetType:CreateResourceSet', createResourceSet_resourceSetType - AWS Resource type of the resources in the ResourceSet

$sel:resourceSetName:CreateResourceSet', createResourceSet_resourceSetName - The name of the ResourceSet to create

$sel:resources:CreateResourceSet', createResourceSet_resources - A list of Resource objects

Request Lenses

createResourceSet_tags :: Lens' CreateResourceSet (Maybe (HashMap Text Text))

Undocumented member.

createResourceSet_resourceSetType :: Lens' CreateResourceSet Text

AWS Resource type of the resources in the ResourceSet

createResourceSet_resourceSetName :: Lens' CreateResourceSet Text

The name of the ResourceSet to create

createResourceSet_resources :: Lens' CreateResourceSet [Resource]

A list of Resource objects

Destructuring the Response

data CreateResourceSetResponse

See: newCreateResourceSetResponse smart constructor.

Constructors

CreateResourceSetResponse'

Fields

resourceSetName :: Maybe Text

The name of the ResourceSet

resourceSetType :: Maybe Text

AWS Resource Type of the resources in the ResourceSet

resources :: Maybe [Resource]

A list of Resource objects

resourceSetArn :: Maybe Text

The arn for the ResourceSet

tags :: Maybe (HashMap Text Text)

httpStatus :: Int

The response's http status code.

newCreateResourceSetResponse

Arguments

:: Int
$sel:httpStatus:CreateResourceSetResponse'

-> CreateResourceSetResponse

Create a value of CreateResourceSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceSetName:CreateResourceSet', createResourceSetResponse_resourceSetName - The name of the ResourceSet

$sel:resourceSetType:CreateResourceSet', createResourceSetResponse_resourceSetType - AWS Resource Type of the resources in the ResourceSet

$sel:resources:CreateResourceSet', createResourceSetResponse_resources - A list of Resource objects

$sel:resourceSetArn:CreateResourceSetResponse', createResourceSetResponse_resourceSetArn - The arn for the ResourceSet

$sel:tags:CreateResourceSet', createResourceSetResponse_tags - Undocumented member.

$sel:httpStatus:CreateResourceSetResponse', createResourceSetResponse_httpStatus - The response's http status code.

Response Lenses

createResourceSetResponse_resourceSetName :: Lens' CreateResourceSetResponse (Maybe Text)

The name of the ResourceSet

createResourceSetResponse_resourceSetType :: Lens' CreateResourceSetResponse (Maybe Text)

AWS Resource Type of the resources in the ResourceSet

createResourceSetResponse_resources :: Lens' CreateResourceSetResponse (Maybe [Resource])

A list of Resource objects

createResourceSetResponse_resourceSetArn :: Lens' CreateResourceSetResponse (Maybe Text)

The arn for the ResourceSet

createResourceSetResponse_tags :: Lens' CreateResourceSetResponse (Maybe (HashMap Text Text))

Undocumented member.

createResourceSetResponse_httpStatus :: Lens' CreateResourceSetResponse Int

The response's http status code.