Amazonka.Route53RecoveryReadiness.GetCell

Description

Returns information about a Cell.

Creating a Request

data GetCell

See: newGetCell smart constructor.

Constructors

GetCell'

Fields

cellName :: Text

The Cell to get

newGetCell

Arguments

:: Text
$sel:cellName:GetCell'

-> GetCell

Create a value of GetCell with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cellName:GetCell', getCell_cellName - The Cell to get

Request Lenses

getCell_cellName :: Lens' GetCell Text

The Cell to get

Destructuring the Response

data GetCellResponse

See: newGetCellResponse smart constructor.

Constructors

GetCellResponse'

Fields

cells :: Maybe [Text]

A list of Cell arns

parentReadinessScopes :: Maybe [Text]

A list of Cell ARNs and/or RecoveryGroup ARNs

cellName :: Maybe Text

The name of the Cell

cellArn :: Maybe Text

The arn for the Cell

tags :: Maybe (HashMap Text Text)

httpStatus :: Int

The response's http status code.

newGetCellResponse

Arguments

:: Int
$sel:httpStatus:GetCellResponse'

-> GetCellResponse

Create a value of GetCellResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cells:GetCellResponse', getCellResponse_cells - A list of Cell arns

$sel:parentReadinessScopes:GetCellResponse', getCellResponse_parentReadinessScopes - A list of Cell ARNs and/or RecoveryGroup ARNs

$sel:cellName:GetCell', getCellResponse_cellName - The name of the Cell

$sel:cellArn:GetCellResponse', getCellResponse_cellArn - The arn for the Cell

$sel:tags:GetCellResponse', getCellResponse_tags - Undocumented member.

$sel:httpStatus:GetCellResponse', getCellResponse_httpStatus - The response's http status code.

Response Lenses

getCellResponse_cells :: Lens' GetCellResponse (Maybe [Text])

A list of Cell arns

getCellResponse_parentReadinessScopes :: Lens' GetCellResponse (Maybe [Text])

A list of Cell ARNs and/or RecoveryGroup ARNs

getCellResponse_cellName :: Lens' GetCellResponse (Maybe Text)

The name of the Cell

getCellResponse_cellArn :: Lens' GetCellResponse (Maybe Text)

The arn for the Cell

getCellResponse_tags :: Lens' GetCellResponse (Maybe (HashMap Text Text))

Undocumented member.

getCellResponse_httpStatus :: Lens' GetCellResponse Int

The response's http status code.