Amazonka.Route53RecoveryReadiness.ListReadinessChecks

Description

Returns a collection of Readiness Checks.

This operation returns paginated results.

Creating a Request

data ListReadinessChecks

See: newListReadinessChecks smart constructor.

Constructors

ListReadinessChecks'

Fields

nextToken :: Maybe Text

A token used to resume pagination from the end of a previous request.

maxResults :: Maybe Natural

Upper bound on number of records to return.

newListReadinessChecks :: ListReadinessChecks

Create a value of ListReadinessChecks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListReadinessChecks', listReadinessChecks_nextToken - A token used to resume pagination from the end of a previous request.

$sel:maxResults:ListReadinessChecks', listReadinessChecks_maxResults - Upper bound on number of records to return.

Request Lenses

listReadinessChecks_nextToken :: Lens' ListReadinessChecks (Maybe Text)

A token used to resume pagination from the end of a previous request.

listReadinessChecks_maxResults :: Lens' ListReadinessChecks (Maybe Natural)

Upper bound on number of records to return.

Destructuring the Response

data ListReadinessChecksResponse

See: newListReadinessChecksResponse smart constructor.

Constructors

ListReadinessChecksResponse'

Fields

readinessChecks :: Maybe [ReadinessCheckOutput]

A list of ReadinessCheck associated with the account

nextToken :: Maybe Text

A token that can be used to resume pagination from the end of the collection.

httpStatus :: Int

The response's http status code.

newListReadinessChecksResponse

Arguments

:: Int
$sel:httpStatus:ListReadinessChecksResponse'

-> ListReadinessChecksResponse

Create a value of ListReadinessChecksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readinessChecks:ListReadinessChecksResponse', listReadinessChecksResponse_readinessChecks - A list of ReadinessCheck associated with the account

$sel:nextToken:ListReadinessChecks', listReadinessChecksResponse_nextToken - A token that can be used to resume pagination from the end of the collection.

$sel:httpStatus:ListReadinessChecksResponse', listReadinessChecksResponse_httpStatus - The response's http status code.

Response Lenses

listReadinessChecksResponse_readinessChecks :: Lens' ListReadinessChecksResponse (Maybe [ReadinessCheckOutput])

A list of ReadinessCheck associated with the account

listReadinessChecksResponse_nextToken :: Lens' ListReadinessChecksResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listReadinessChecksResponse_httpStatus :: Lens' ListReadinessChecksResponse Int

The response's http status code.