Amazonka.Route53Resolver.AssociateResolverEndpointIpAddress

Description

Adds IP addresses to an inbound or an outbound Resolver endpoint. If you want to add more than one IP address, submit one AssociateResolverEndpointIpAddress request for each IP address.

To remove an IP address from an endpoint, see DisassociateResolverEndpointIpAddress.

Creating a Request

data AssociateResolverEndpointIpAddress

See: newAssociateResolverEndpointIpAddress smart constructor.

Constructors

AssociateResolverEndpointIpAddress'

Fields

resolverEndpointId :: Text

The ID of the Resolver endpoint that you want to associate IP addresses with.

ipAddress :: IpAddressUpdate

Either the IPv4 address that you want to add to a Resolver endpoint or a subnet ID. If you specify a subnet ID, Resolver chooses an IP address for you from the available IPs in the specified subnet.

newAssociateResolverEndpointIpAddress

Arguments

:: Text
$sel:resolverEndpointId:AssociateResolverEndpointIpAddress'

-> IpAddressUpdate
$sel:ipAddress:AssociateResolverEndpointIpAddress'

-> AssociateResolverEndpointIpAddress

Create a value of AssociateResolverEndpointIpAddress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverEndpointId:AssociateResolverEndpointIpAddress', associateResolverEndpointIpAddress_resolverEndpointId - The ID of the Resolver endpoint that you want to associate IP addresses with.

$sel:ipAddress:AssociateResolverEndpointIpAddress', associateResolverEndpointIpAddress_ipAddress - Either the IPv4 address that you want to add to a Resolver endpoint or a subnet ID. If you specify a subnet ID, Resolver chooses an IP address for you from the available IPs in the specified subnet.

Request Lenses

associateResolverEndpointIpAddress_resolverEndpointId :: Lens' AssociateResolverEndpointIpAddress Text

The ID of the Resolver endpoint that you want to associate IP addresses with.

associateResolverEndpointIpAddress_ipAddress :: Lens' AssociateResolverEndpointIpAddress IpAddressUpdate

Either the IPv4 address that you want to add to a Resolver endpoint or a subnet ID. If you specify a subnet ID, Resolver chooses an IP address for you from the available IPs in the specified subnet.

Destructuring the Response

data AssociateResolverEndpointIpAddressResponse

See: newAssociateResolverEndpointIpAddressResponse smart constructor.

Constructors

AssociateResolverEndpointIpAddressResponse'

Fields

resolverEndpoint :: Maybe ResolverEndpoint

The response to an AssociateResolverEndpointIpAddress request.

httpStatus :: Int

The response's http status code.

newAssociateResolverEndpointIpAddressResponse

Arguments

:: Int
$sel:httpStatus:AssociateResolverEndpointIpAddressResponse'

-> AssociateResolverEndpointIpAddressResponse

Create a value of AssociateResolverEndpointIpAddressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverEndpoint:AssociateResolverEndpointIpAddressResponse', associateResolverEndpointIpAddressResponse_resolverEndpoint - The response to an AssociateResolverEndpointIpAddress request.

$sel:httpStatus:AssociateResolverEndpointIpAddressResponse', associateResolverEndpointIpAddressResponse_httpStatus - The response's http status code.

Response Lenses

associateResolverEndpointIpAddressResponse_resolverEndpoint :: Lens' AssociateResolverEndpointIpAddressResponse (Maybe ResolverEndpoint)

The response to an AssociateResolverEndpointIpAddress request.

associateResolverEndpointIpAddressResponse_httpStatus :: Lens' AssociateResolverEndpointIpAddressResponse Int

The response's http status code.