Amazonka.Route53Resolver.CreateFirewallRule

Description

Creates a single DNS Firewall rule in the specified rule group, using the specified domain list.

Creating a Request

data CreateFirewallRule

See: newCreateFirewallRule smart constructor.

Constructors

CreateFirewallRule'

Fields

blockOverrideDnsType :: Maybe BlockOverrideDnsType

The DNS record's type. This determines the format of the record value that you provided in BlockOverrideDomain. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

This setting is required if the BlockResponse setting is OVERRIDE.

blockResponse :: Maybe BlockResponse

The way that you want DNS Firewall to block the request, used with the rule action setting BLOCK.

NODATA - Respond indicating that the query was successful, but no response is available for it.

NXDOMAIN - Respond indicating that the domain name that's in the query doesn't exist.

OVERRIDE - Provide a custom override in the response. This option requires custom handling details in the rule's BlockOverride* settings.

This setting is required if the rule action setting is BLOCK.

blockOverrideTtl :: Maybe Natural

The recommended amount of time, in seconds, for the DNS resolver or web browser to cache the provided override record. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

This setting is required if the BlockResponse setting is OVERRIDE.

blockOverrideDomain :: Maybe Text

The custom DNS record to send back in response to the query. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

This setting is required if the BlockResponse setting is OVERRIDE.

creatorRequestId :: Text

A unique string that identifies the request and that allows you to retry failed requests without the risk of running the operation twice. CreatorRequestId can be any unique string, for example, a date/time stamp.

firewallRuleGroupId :: Text

The unique identifier of the firewall rule group where you want to create the rule.

firewallDomainListId :: Text

The ID of the domain list that you want to use in the rule.

priority :: Int

The setting that determines the processing order of the rule in the rule group. DNS Firewall processes the rules in a rule group by order of priority, starting from the lowest setting.

You must specify a unique priority for each rule in a rule group. To make it easier to insert rules later, leave space between the numbers, for example, use 100, 200, and so on. You can change the priority setting for the rules in a rule group at any time.

action :: Action

The action that DNS Firewall should take on a DNS query when it matches one of the domains in the rule's domain list:

ALLOW - Permit the request to go through.

ALERT - Permit the request and send metrics and logs to Cloud Watch.

BLOCK - Disallow the request. This option requires additional details in the rule's BlockResponse.

name :: Text

A name that lets you identify the rule in the rule group.

newCreateFirewallRule

Arguments

:: Text
$sel:creatorRequestId:CreateFirewallRule'

-> Text
$sel:firewallRuleGroupId:CreateFirewallRule'

-> Text
$sel:firewallDomainListId:CreateFirewallRule'

-> Int
$sel:priority:CreateFirewallRule'

-> Action
$sel:action:CreateFirewallRule'

-> Text
$sel:name:CreateFirewallRule'

-> CreateFirewallRule

Create a value of CreateFirewallRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockOverrideDnsType:CreateFirewallRule', createFirewallRule_blockOverrideDnsType - The DNS record's type. This determines the format of the record value that you provided in BlockOverrideDomain. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

This setting is required if the BlockResponse setting is OVERRIDE.

$sel:blockResponse:CreateFirewallRule', createFirewallRule_blockResponse - The way that you want DNS Firewall to block the request, used with the rule action setting BLOCK.

NODATA - Respond indicating that the query was successful, but no response is available for it.

NXDOMAIN - Respond indicating that the domain name that's in the query doesn't exist.

OVERRIDE - Provide a custom override in the response. This option requires custom handling details in the rule's BlockOverride* settings.

This setting is required if the rule action setting is BLOCK.

$sel:blockOverrideTtl:CreateFirewallRule', createFirewallRule_blockOverrideTtl - The recommended amount of time, in seconds, for the DNS resolver or web browser to cache the provided override record. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

This setting is required if the BlockResponse setting is OVERRIDE.

$sel:blockOverrideDomain:CreateFirewallRule', createFirewallRule_blockOverrideDomain - The custom DNS record to send back in response to the query. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

This setting is required if the BlockResponse setting is OVERRIDE.

$sel:creatorRequestId:CreateFirewallRule', createFirewallRule_creatorRequestId - A unique string that identifies the request and that allows you to retry failed requests without the risk of running the operation twice. CreatorRequestId can be any unique string, for example, a date/time stamp.

$sel:firewallRuleGroupId:CreateFirewallRule', createFirewallRule_firewallRuleGroupId - The unique identifier of the firewall rule group where you want to create the rule.

$sel:firewallDomainListId:CreateFirewallRule', createFirewallRule_firewallDomainListId - The ID of the domain list that you want to use in the rule.

$sel:priority:CreateFirewallRule', createFirewallRule_priority - The setting that determines the processing order of the rule in the rule group. DNS Firewall processes the rules in a rule group by order of priority, starting from the lowest setting.

You must specify a unique priority for each rule in a rule group. To make it easier to insert rules later, leave space between the numbers, for example, use 100, 200, and so on. You can change the priority setting for the rules in a rule group at any time.

$sel:action:CreateFirewallRule', createFirewallRule_action - The action that DNS Firewall should take on a DNS query when it matches one of the domains in the rule's domain list:

ALLOW - Permit the request to go through.

ALERT - Permit the request and send metrics and logs to Cloud Watch.

BLOCK - Disallow the request. This option requires additional details in the rule's BlockResponse.

$sel:name:CreateFirewallRule', createFirewallRule_name - A name that lets you identify the rule in the rule group.

Request Lenses

createFirewallRule_blockOverrideDnsType :: Lens' CreateFirewallRule (Maybe BlockOverrideDnsType)

The DNS record's type. This determines the format of the record value that you provided in BlockOverrideDomain. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

This setting is required if the BlockResponse setting is OVERRIDE.

createFirewallRule_blockResponse :: Lens' CreateFirewallRule (Maybe BlockResponse)

The way that you want DNS Firewall to block the request, used with the rule action setting BLOCK.

NODATA - Respond indicating that the query was successful, but no response is available for it.

NXDOMAIN - Respond indicating that the domain name that's in the query doesn't exist.

OVERRIDE - Provide a custom override in the response. This option requires custom handling details in the rule's BlockOverride* settings.

This setting is required if the rule action setting is BLOCK.

createFirewallRule_blockOverrideTtl :: Lens' CreateFirewallRule (Maybe Natural)

The recommended amount of time, in seconds, for the DNS resolver or web browser to cache the provided override record. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

This setting is required if the BlockResponse setting is OVERRIDE.

createFirewallRule_blockOverrideDomain :: Lens' CreateFirewallRule (Maybe Text)

The custom DNS record to send back in response to the query. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

This setting is required if the BlockResponse setting is OVERRIDE.

createFirewallRule_creatorRequestId :: Lens' CreateFirewallRule Text

A unique string that identifies the request and that allows you to retry failed requests without the risk of running the operation twice. CreatorRequestId can be any unique string, for example, a date/time stamp.

createFirewallRule_firewallRuleGroupId :: Lens' CreateFirewallRule Text

The unique identifier of the firewall rule group where you want to create the rule.

createFirewallRule_firewallDomainListId :: Lens' CreateFirewallRule Text

The ID of the domain list that you want to use in the rule.

createFirewallRule_priority :: Lens' CreateFirewallRule Int

The setting that determines the processing order of the rule in the rule group. DNS Firewall processes the rules in a rule group by order of priority, starting from the lowest setting.

You must specify a unique priority for each rule in a rule group. To make it easier to insert rules later, leave space between the numbers, for example, use 100, 200, and so on. You can change the priority setting for the rules in a rule group at any time.

createFirewallRule_action :: Lens' CreateFirewallRule Action

The action that DNS Firewall should take on a DNS query when it matches one of the domains in the rule's domain list:

ALLOW - Permit the request to go through.

ALERT - Permit the request and send metrics and logs to Cloud Watch.

BLOCK - Disallow the request. This option requires additional details in the rule's BlockResponse.

createFirewallRule_name :: Lens' CreateFirewallRule Text

A name that lets you identify the rule in the rule group.

Destructuring the Response

data CreateFirewallRuleResponse

See: newCreateFirewallRuleResponse smart constructor.

Constructors

CreateFirewallRuleResponse'

Fields

firewallRule :: Maybe FirewallRule

The firewall rule that you just created.

httpStatus :: Int

The response's http status code.

newCreateFirewallRuleResponse

Arguments

:: Int
$sel:httpStatus:CreateFirewallRuleResponse'

-> CreateFirewallRuleResponse

Create a value of CreateFirewallRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallRule:CreateFirewallRuleResponse', createFirewallRuleResponse_firewallRule - The firewall rule that you just created.

$sel:httpStatus:CreateFirewallRuleResponse', createFirewallRuleResponse_httpStatus - The response's http status code.

Response Lenses

createFirewallRuleResponse_firewallRule :: Lens' CreateFirewallRuleResponse (Maybe FirewallRule)

The firewall rule that you just created.

createFirewallRuleResponse_httpStatus :: Lens' CreateFirewallRuleResponse Int

The response's http status code.