Amazonka.Route53Resolver.CreateResolverQueryLogConfig

Description

Creates a Resolver query logging configuration, which defines where you want Resolver to save DNS query logs that originate in your VPCs. Resolver can log queries only for VPCs that are in the same Region as the query logging configuration.

To specify which VPCs you want to log queries for, you use AssociateResolverQueryLogConfig. For more information, see AssociateResolverQueryLogConfig.

You can optionally use Resource Access Manager (RAM) to share a query logging configuration with other Amazon Web Services accounts. The other accounts can then associate VPCs with the configuration. The query logs that Resolver creates for a configuration include all DNS queries that originate in all VPCs that are associated with the configuration.

Creating a Request

data CreateResolverQueryLogConfig

See: newCreateResolverQueryLogConfig smart constructor.

Constructors

CreateResolverQueryLogConfig'

Fields

tags :: Maybe [Tag]

A list of the tag keys and values that you want to associate with the query logging configuration.

name :: Text

The name that you want to give the query logging configuration.

destinationArn :: Text

The ARN of the resource that you want Resolver to send query logs. You can send query logs to an S3 bucket, a CloudWatch Logs log group, or a Kinesis Data Firehose delivery stream. Examples of valid values include the following:

S3 bucket:

arn:aws:s3:::examplebucket

You can optionally append a file prefix to the end of the ARN.

arn:aws:s3:::examplebucket/development/

CloudWatch Logs log group:

arn:aws:logs:us-west-1:123456789012:log-group:/mystack-testgroup-12ABC1AB12A1:*

Kinesis Data Firehose delivery stream:

arn:aws:kinesis:us-east-2:0123456789:stream/my_stream_name

creatorRequestId :: Text

A unique string that identifies the request and that allows failed requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string, for example, a date/time stamp.

newCreateResolverQueryLogConfig

Arguments

:: Text
$sel:name:CreateResolverQueryLogConfig'

-> Text
$sel:destinationArn:CreateResolverQueryLogConfig'

-> Text
$sel:creatorRequestId:CreateResolverQueryLogConfig'

-> CreateResolverQueryLogConfig

Create a value of CreateResolverQueryLogConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateResolverQueryLogConfig', createResolverQueryLogConfig_tags - A list of the tag keys and values that you want to associate with the query logging configuration.

$sel:name:CreateResolverQueryLogConfig', createResolverQueryLogConfig_name - The name that you want to give the query logging configuration.

$sel:destinationArn:CreateResolverQueryLogConfig', createResolverQueryLogConfig_destinationArn - The ARN of the resource that you want Resolver to send query logs. You can send query logs to an S3 bucket, a CloudWatch Logs log group, or a Kinesis Data Firehose delivery stream. Examples of valid values include the following:

S3 bucket:

arn:aws:s3:::examplebucket

You can optionally append a file prefix to the end of the ARN.

arn:aws:s3:::examplebucket/development/

CloudWatch Logs log group:

arn:aws:logs:us-west-1:123456789012:log-group:/mystack-testgroup-12ABC1AB12A1:*

Kinesis Data Firehose delivery stream:

arn:aws:kinesis:us-east-2:0123456789:stream/my_stream_name

$sel:creatorRequestId:CreateResolverQueryLogConfig', createResolverQueryLogConfig_creatorRequestId - A unique string that identifies the request and that allows failed requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string, for example, a date/time stamp.

Request Lenses

createResolverQueryLogConfig_tags :: Lens' CreateResolverQueryLogConfig (Maybe [Tag])

A list of the tag keys and values that you want to associate with the query logging configuration.

createResolverQueryLogConfig_name :: Lens' CreateResolverQueryLogConfig Text

The name that you want to give the query logging configuration.

createResolverQueryLogConfig_destinationArn :: Lens' CreateResolverQueryLogConfig Text

The ARN of the resource that you want Resolver to send query logs. You can send query logs to an S3 bucket, a CloudWatch Logs log group, or a Kinesis Data Firehose delivery stream. Examples of valid values include the following:

S3 bucket:

arn:aws:s3:::examplebucket

You can optionally append a file prefix to the end of the ARN.

arn:aws:s3:::examplebucket/development/

CloudWatch Logs log group:

arn:aws:logs:us-west-1:123456789012:log-group:/mystack-testgroup-12ABC1AB12A1:*

Kinesis Data Firehose delivery stream:

arn:aws:kinesis:us-east-2:0123456789:stream/my_stream_name

createResolverQueryLogConfig_creatorRequestId :: Lens' CreateResolverQueryLogConfig Text

A unique string that identifies the request and that allows failed requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string, for example, a date/time stamp.

Destructuring the Response

data CreateResolverQueryLogConfigResponse

See: newCreateResolverQueryLogConfigResponse smart constructor.

Constructors

CreateResolverQueryLogConfigResponse'

Fields

resolverQueryLogConfig :: Maybe ResolverQueryLogConfig

Information about the CreateResolverQueryLogConfig request, including the status of the request.

httpStatus :: Int

The response's http status code.

newCreateResolverQueryLogConfigResponse

Arguments

:: Int
$sel:httpStatus:CreateResolverQueryLogConfigResponse'

-> CreateResolverQueryLogConfigResponse

Create a value of CreateResolverQueryLogConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverQueryLogConfig:CreateResolverQueryLogConfigResponse', createResolverQueryLogConfigResponse_resolverQueryLogConfig - Information about the CreateResolverQueryLogConfig request, including the status of the request.

$sel:httpStatus:CreateResolverQueryLogConfigResponse', createResolverQueryLogConfigResponse_httpStatus - The response's http status code.

Response Lenses

createResolverQueryLogConfigResponse_resolverQueryLogConfig :: Lens' CreateResolverQueryLogConfigResponse (Maybe ResolverQueryLogConfig)

Information about the CreateResolverQueryLogConfig request, including the status of the request.

createResolverQueryLogConfigResponse_httpStatus :: Lens' CreateResolverQueryLogConfigResponse Int

The response's http status code.