Amazonka.Route53Resolver.CreateResolverRule

Description

For DNS queries that originate in your VPCs, specifies which Resolver endpoint the queries pass through, one domain name that you want to forward to your network, and the IP addresses of the DNS resolvers in your network.

Creating a Request

data CreateResolverRule

See: newCreateResolverRule smart constructor.

Constructors

CreateResolverRule'

Fields

resolverEndpointId :: Maybe Text

The ID of the outbound Resolver endpoint that you want to use to route DNS queries to the IP addresses that you specify in TargetIps.

targetIps :: Maybe (NonEmpty TargetAddress)

The IPs that you want Resolver to forward DNS queries to. You can specify only IPv4 addresses. Separate IP addresses with a space.

TargetIps is available only when the value of Rule type is FORWARD.

name :: Maybe Text

A friendly name that lets you easily find a rule in the Resolver dashboard in the Route 53 console.

tags :: Maybe [Tag]

A list of the tag keys and values that you want to associate with the endpoint.

creatorRequestId :: Text

A unique string that identifies the request and that allows failed requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string, for example, a date/time stamp.

ruleType :: RuleTypeOption

When you want to forward DNS queries for specified domain name to resolvers on your network, specify FORWARD.

When you have a forwarding rule to forward DNS queries for a domain to your network and you want Resolver to process queries for a subdomain of that domain, specify SYSTEM.

For example, to forward DNS queries for example.com to resolvers on your network, you create a rule and specify FORWARD for RuleType. To then have Resolver process queries for apex.example.com, you create a rule and specify SYSTEM for RuleType.

Currently, only Resolver can create rules that have a value of RECURSIVE for RuleType.

domainName :: Text

DNS queries for this domain name are forwarded to the IP addresses that you specify in TargetIps. If a query matches multiple Resolver rules (example.com and www.example.com), outbound DNS queries are routed using the Resolver rule that contains the most specific domain name (www.example.com).

newCreateResolverRule

Arguments

:: Text
$sel:creatorRequestId:CreateResolverRule'

-> RuleTypeOption
$sel:ruleType:CreateResolverRule'

-> Text
$sel:domainName:CreateResolverRule'

-> CreateResolverRule

Create a value of CreateResolverRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverEndpointId:CreateResolverRule', createResolverRule_resolverEndpointId - The ID of the outbound Resolver endpoint that you want to use to route DNS queries to the IP addresses that you specify in TargetIps.

$sel:targetIps:CreateResolverRule', createResolverRule_targetIps - The IPs that you want Resolver to forward DNS queries to. You can specify only IPv4 addresses. Separate IP addresses with a space.

TargetIps is available only when the value of Rule type is FORWARD.

$sel:name:CreateResolverRule', createResolverRule_name - A friendly name that lets you easily find a rule in the Resolver dashboard in the Route 53 console.

$sel:tags:CreateResolverRule', createResolverRule_tags - A list of the tag keys and values that you want to associate with the endpoint.

$sel:creatorRequestId:CreateResolverRule', createResolverRule_creatorRequestId - A unique string that identifies the request and that allows failed requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string, for example, a date/time stamp.

$sel:ruleType:CreateResolverRule', createResolverRule_ruleType - When you want to forward DNS queries for specified domain name to resolvers on your network, specify FORWARD.

When you have a forwarding rule to forward DNS queries for a domain to your network and you want Resolver to process queries for a subdomain of that domain, specify SYSTEM.

For example, to forward DNS queries for example.com to resolvers on your network, you create a rule and specify FORWARD for RuleType. To then have Resolver process queries for apex.example.com, you create a rule and specify SYSTEM for RuleType.

Currently, only Resolver can create rules that have a value of RECURSIVE for RuleType.

$sel:domainName:CreateResolverRule', createResolverRule_domainName - DNS queries for this domain name are forwarded to the IP addresses that you specify in TargetIps. If a query matches multiple Resolver rules (example.com and www.example.com), outbound DNS queries are routed using the Resolver rule that contains the most specific domain name (www.example.com).

Request Lenses

createResolverRule_resolverEndpointId :: Lens' CreateResolverRule (Maybe Text)

The ID of the outbound Resolver endpoint that you want to use to route DNS queries to the IP addresses that you specify in TargetIps.

createResolverRule_targetIps :: Lens' CreateResolverRule (Maybe (NonEmpty TargetAddress))

The IPs that you want Resolver to forward DNS queries to. You can specify only IPv4 addresses. Separate IP addresses with a space.

TargetIps is available only when the value of Rule type is FORWARD.

createResolverRule_name :: Lens' CreateResolverRule (Maybe Text)

A friendly name that lets you easily find a rule in the Resolver dashboard in the Route 53 console.

createResolverRule_tags :: Lens' CreateResolverRule (Maybe [Tag])

A list of the tag keys and values that you want to associate with the endpoint.

createResolverRule_creatorRequestId :: Lens' CreateResolverRule Text

A unique string that identifies the request and that allows failed requests to be retried without the risk of running the operation twice. CreatorRequestId can be any unique string, for example, a date/time stamp.

createResolverRule_ruleType :: Lens' CreateResolverRule RuleTypeOption

When you want to forward DNS queries for specified domain name to resolvers on your network, specify FORWARD.

When you have a forwarding rule to forward DNS queries for a domain to your network and you want Resolver to process queries for a subdomain of that domain, specify SYSTEM.

For example, to forward DNS queries for example.com to resolvers on your network, you create a rule and specify FORWARD for RuleType. To then have Resolver process queries for apex.example.com, you create a rule and specify SYSTEM for RuleType.

Currently, only Resolver can create rules that have a value of RECURSIVE for RuleType.

createResolverRule_domainName :: Lens' CreateResolverRule Text

DNS queries for this domain name are forwarded to the IP addresses that you specify in TargetIps. If a query matches multiple Resolver rules (example.com and www.example.com), outbound DNS queries are routed using the Resolver rule that contains the most specific domain name (www.example.com).

Destructuring the Response

data CreateResolverRuleResponse

See: newCreateResolverRuleResponse smart constructor.

Constructors

CreateResolverRuleResponse'

Fields

resolverRule :: Maybe ResolverRule

Information about the CreateResolverRule request, including the status of the request.

httpStatus :: Int

The response's http status code.

newCreateResolverRuleResponse

Arguments

:: Int
$sel:httpStatus:CreateResolverRuleResponse'

-> CreateResolverRuleResponse

Create a value of CreateResolverRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverRule:CreateResolverRuleResponse', createResolverRuleResponse_resolverRule - Information about the CreateResolverRule request, including the status of the request.

$sel:httpStatus:CreateResolverRuleResponse', createResolverRuleResponse_httpStatus - The response's http status code.

Response Lenses

createResolverRuleResponse_resolverRule :: Lens' CreateResolverRuleResponse (Maybe ResolverRule)

Information about the CreateResolverRule request, including the status of the request.

createResolverRuleResponse_httpStatus :: Lens' CreateResolverRuleResponse Int

The response's http status code.