Amazonka.Route53Resolver.ListResolverEndpointIpAddresses

Description

Gets the IP addresses for a specified Resolver endpoint.

This operation returns paginated results.

Creating a Request

data ListResolverEndpointIpAddresses

See: newListResolverEndpointIpAddresses smart constructor.

Constructors

ListResolverEndpointIpAddresses'

Fields

nextToken :: Maybe Text

For the first ListResolverEndpointIpAddresses request, omit this value.

If the specified Resolver endpoint has more than MaxResults IP addresses, you can submit another ListResolverEndpointIpAddresses request to get the next group of IP addresses. In the next request, specify the value of NextToken from the previous response.

maxResults :: Maybe Natural

The maximum number of IP addresses that you want to return in the response to a ListResolverEndpointIpAddresses request. If you don't specify a value for MaxResults, Resolver returns up to 100 IP addresses.

resolverEndpointId :: Text

The ID of the Resolver endpoint that you want to get IP addresses for.

newListResolverEndpointIpAddresses

Arguments

:: Text
$sel:resolverEndpointId:ListResolverEndpointIpAddresses'

-> ListResolverEndpointIpAddresses

Create a value of ListResolverEndpointIpAddresses with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListResolverEndpointIpAddresses', listResolverEndpointIpAddresses_nextToken - For the first ListResolverEndpointIpAddresses request, omit this value.

If the specified Resolver endpoint has more than MaxResults IP addresses, you can submit another ListResolverEndpointIpAddresses request to get the next group of IP addresses. In the next request, specify the value of NextToken from the previous response.

$sel:maxResults:ListResolverEndpointIpAddresses', listResolverEndpointIpAddresses_maxResults - The maximum number of IP addresses that you want to return in the response to a ListResolverEndpointIpAddresses request. If you don't specify a value for MaxResults, Resolver returns up to 100 IP addresses.

$sel:resolverEndpointId:ListResolverEndpointIpAddresses', listResolverEndpointIpAddresses_resolverEndpointId - The ID of the Resolver endpoint that you want to get IP addresses for.

Request Lenses

listResolverEndpointIpAddresses_nextToken :: Lens' ListResolverEndpointIpAddresses (Maybe Text)

For the first ListResolverEndpointIpAddresses request, omit this value.

If the specified Resolver endpoint has more than MaxResults IP addresses, you can submit another ListResolverEndpointIpAddresses request to get the next group of IP addresses. In the next request, specify the value of NextToken from the previous response.

listResolverEndpointIpAddresses_maxResults :: Lens' ListResolverEndpointIpAddresses (Maybe Natural)

The maximum number of IP addresses that you want to return in the response to a ListResolverEndpointIpAddresses request. If you don't specify a value for MaxResults, Resolver returns up to 100 IP addresses.

listResolverEndpointIpAddresses_resolverEndpointId :: Lens' ListResolverEndpointIpAddresses Text

The ID of the Resolver endpoint that you want to get IP addresses for.

Destructuring the Response

data ListResolverEndpointIpAddressesResponse

See: newListResolverEndpointIpAddressesResponse smart constructor.

Constructors

ListResolverEndpointIpAddressesResponse'

Fields

nextToken :: Maybe Text

If the specified endpoint has more than MaxResults IP addresses, you can submit another ListResolverEndpointIpAddresses request to get the next group of IP addresses. In the next request, specify the value of NextToken from the previous response.

maxResults :: Maybe Natural

The value that you specified for MaxResults in the request.

ipAddresses :: Maybe [IpAddressResponse]

Information about the IP addresses in your VPC that DNS queries originate from (for outbound endpoints) or that you forward DNS queries to (for inbound endpoints).

httpStatus :: Int

The response's http status code.

newListResolverEndpointIpAddressesResponse

Arguments

:: Int
$sel:httpStatus:ListResolverEndpointIpAddressesResponse'

-> ListResolverEndpointIpAddressesResponse

Create a value of ListResolverEndpointIpAddressesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListResolverEndpointIpAddresses', listResolverEndpointIpAddressesResponse_nextToken - If the specified endpoint has more than MaxResults IP addresses, you can submit another ListResolverEndpointIpAddresses request to get the next group of IP addresses. In the next request, specify the value of NextToken from the previous response.

$sel:maxResults:ListResolverEndpointIpAddresses', listResolverEndpointIpAddressesResponse_maxResults - The value that you specified for MaxResults in the request.

$sel:ipAddresses:ListResolverEndpointIpAddressesResponse', listResolverEndpointIpAddressesResponse_ipAddresses - Information about the IP addresses in your VPC that DNS queries originate from (for outbound endpoints) or that you forward DNS queries to (for inbound endpoints).

$sel:httpStatus:ListResolverEndpointIpAddressesResponse', listResolverEndpointIpAddressesResponse_httpStatus - The response's http status code.

Response Lenses

listResolverEndpointIpAddressesResponse_nextToken :: Lens' ListResolverEndpointIpAddressesResponse (Maybe Text)

If the specified endpoint has more than MaxResults IP addresses, you can submit another ListResolverEndpointIpAddresses request to get the next group of IP addresses. In the next request, specify the value of NextToken from the previous response.

listResolverEndpointIpAddressesResponse_maxResults :: Lens' ListResolverEndpointIpAddressesResponse (Maybe Natural)

The value that you specified for MaxResults in the request.

listResolverEndpointIpAddressesResponse_ipAddresses :: Lens' ListResolverEndpointIpAddressesResponse (Maybe [IpAddressResponse])

Information about the IP addresses in your VPC that DNS queries originate from (for outbound endpoints) or that you forward DNS queries to (for inbound endpoints).

listResolverEndpointIpAddressesResponse_httpStatus :: Lens' ListResolverEndpointIpAddressesResponse Int

The response's http status code.