Amazonka.Route53Resolver.ListResolverEndpoints

Description

Lists all the Resolver endpoints that were created using the current Amazon Web Services account.

This operation returns paginated results.

Creating a Request

data ListResolverEndpoints

See: newListResolverEndpoints smart constructor.

Constructors

ListResolverEndpoints'

Fields

filters :: Maybe [Filter]

An optional specification to return a subset of Resolver endpoints, such as all inbound Resolver endpoints.

If you submit a second or subsequent ListResolverEndpoints request and specify the NextToken parameter, you must use the same values for Filters, if any, as in the previous request.

nextToken :: Maybe Text

For the first ListResolverEndpoints request, omit this value.

If you have more than MaxResults Resolver endpoints, you can submit another ListResolverEndpoints request to get the next group of Resolver endpoints. In the next request, specify the value of NextToken from the previous response.

maxResults :: Maybe Natural

The maximum number of Resolver endpoints that you want to return in the response to a ListResolverEndpoints request. If you don't specify a value for MaxResults, Resolver returns up to 100 Resolver endpoints.

newListResolverEndpoints :: ListResolverEndpoints

Create a value of ListResolverEndpoints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListResolverEndpoints', listResolverEndpoints_filters - An optional specification to return a subset of Resolver endpoints, such as all inbound Resolver endpoints.

If you submit a second or subsequent ListResolverEndpoints request and specify the NextToken parameter, you must use the same values for Filters, if any, as in the previous request.

$sel:nextToken:ListResolverEndpoints', listResolverEndpoints_nextToken - For the first ListResolverEndpoints request, omit this value.

If you have more than MaxResults Resolver endpoints, you can submit another ListResolverEndpoints request to get the next group of Resolver endpoints. In the next request, specify the value of NextToken from the previous response.

$sel:maxResults:ListResolverEndpoints', listResolverEndpoints_maxResults - The maximum number of Resolver endpoints that you want to return in the response to a ListResolverEndpoints request. If you don't specify a value for MaxResults, Resolver returns up to 100 Resolver endpoints.

Request Lenses

listResolverEndpoints_filters :: Lens' ListResolverEndpoints (Maybe [Filter])

An optional specification to return a subset of Resolver endpoints, such as all inbound Resolver endpoints.

If you submit a second or subsequent ListResolverEndpoints request and specify the NextToken parameter, you must use the same values for Filters, if any, as in the previous request.

listResolverEndpoints_nextToken :: Lens' ListResolverEndpoints (Maybe Text)

For the first ListResolverEndpoints request, omit this value.

If you have more than MaxResults Resolver endpoints, you can submit another ListResolverEndpoints request to get the next group of Resolver endpoints. In the next request, specify the value of NextToken from the previous response.

listResolverEndpoints_maxResults :: Lens' ListResolverEndpoints (Maybe Natural)

The maximum number of Resolver endpoints that you want to return in the response to a ListResolverEndpoints request. If you don't specify a value for MaxResults, Resolver returns up to 100 Resolver endpoints.

Destructuring the Response

data ListResolverEndpointsResponse

See: newListResolverEndpointsResponse smart constructor.

Constructors

ListResolverEndpointsResponse'

Fields

resolverEndpoints :: Maybe [ResolverEndpoint]

The Resolver endpoints that were created by using the current Amazon Web Services account, and that match the specified filters, if any.

nextToken :: Maybe Text

If more than MaxResults IP addresses match the specified criteria, you can submit another ListResolverEndpoint request to get the next group of results. In the next request, specify the value of NextToken from the previous response.

maxResults :: Maybe Natural

The value that you specified for MaxResults in the request.

httpStatus :: Int

The response's http status code.

newListResolverEndpointsResponse

Arguments

:: Int
$sel:httpStatus:ListResolverEndpointsResponse'

-> ListResolverEndpointsResponse

Create a value of ListResolverEndpointsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverEndpoints:ListResolverEndpointsResponse', listResolverEndpointsResponse_resolverEndpoints - The Resolver endpoints that were created by using the current Amazon Web Services account, and that match the specified filters, if any.

$sel:nextToken:ListResolverEndpoints', listResolverEndpointsResponse_nextToken - If more than MaxResults IP addresses match the specified criteria, you can submit another ListResolverEndpoint request to get the next group of results. In the next request, specify the value of NextToken from the previous response.

$sel:maxResults:ListResolverEndpoints', listResolverEndpointsResponse_maxResults - The value that you specified for MaxResults in the request.

$sel:httpStatus:ListResolverEndpointsResponse', listResolverEndpointsResponse_httpStatus - The response's http status code.

Response Lenses

listResolverEndpointsResponse_resolverEndpoints :: Lens' ListResolverEndpointsResponse (Maybe [ResolverEndpoint])

The Resolver endpoints that were created by using the current Amazon Web Services account, and that match the specified filters, if any.

listResolverEndpointsResponse_nextToken :: Lens' ListResolverEndpointsResponse (Maybe Text)

If more than MaxResults IP addresses match the specified criteria, you can submit another ListResolverEndpoint request to get the next group of results. In the next request, specify the value of NextToken from the previous response.

listResolverEndpointsResponse_maxResults :: Lens' ListResolverEndpointsResponse (Maybe Natural)

The value that you specified for MaxResults in the request.

listResolverEndpointsResponse_httpStatus :: Lens' ListResolverEndpointsResponse Int

The response's http status code.