Amazonka.Route53Resolver.ListResolverQueryLogConfigAssociations

Description

Lists information about associations between Amazon VPCs and query logging configurations.

This operation returns paginated results.

Creating a Request

data ListResolverQueryLogConfigAssociations

See: newListResolverQueryLogConfigAssociations smart constructor.

Constructors

ListResolverQueryLogConfigAssociations'

Fields

filters :: Maybe [Filter]

An optional specification to return a subset of query logging associations.

If you submit a second or subsequent ListResolverQueryLogConfigAssociations request and specify the NextToken parameter, you must use the same values for Filters, if any, as in the previous request.

nextToken :: Maybe Text

For the first ListResolverQueryLogConfigAssociations request, omit this value.

If there are more than MaxResults query logging associations that match the values that you specify for Filters, you can submit another ListResolverQueryLogConfigAssociations request to get the next group of associations. In the next request, specify the value of NextToken from the previous response.

sortOrder :: Maybe SortOrder

If you specified a value for SortBy, the order that you want query logging associations to be listed in, ASCENDING or DESCENDING.

If you submit a second or subsequent ListResolverQueryLogConfigAssociations request and specify the NextToken parameter, you must use the same value for SortOrder, if any, as in the previous request.

maxResults :: Maybe Natural

The maximum number of query logging associations that you want to return in the response to a ListResolverQueryLogConfigAssociations request. If you don't specify a value for MaxResults, Resolver returns up to 100 query logging associations.

sortBy :: Maybe Text

The element that you want Resolver to sort query logging associations by.

If you submit a second or subsequent ListResolverQueryLogConfigAssociations request and specify the NextToken parameter, you must use the same value for SortBy, if any, as in the previous request.

Valid values include the following elements:

CreationTime: The ID of the query logging association.

Error: If the value of Status is FAILED, the value of Error indicates the cause:

DESTINATION_NOT_FOUND: The specified destination (for example, an Amazon S3 bucket) was deleted.

ACCESS_DENIED: Permissions don't allow sending logs to the destination.

If Status is a value other than FAILED, ERROR is null.

Id: The ID of the query logging association

ResolverQueryLogConfigId: The ID of the query logging configuration

ResourceId: The ID of the VPC that is associated with the query logging configuration

Status: The current status of the configuration. Valid values include the following:

CREATING: Resolver is creating an association between an Amazon VPC and a query logging configuration.

CREATED: The association between an Amazon VPC and a query logging configuration was successfully created. Resolver is logging queries that originate in the specified VPC.

DELETING: Resolver is deleting this query logging association.

FAILED: Resolver either couldn't create or couldn't delete the query logging association. Here are two common causes:

The specified destination (for example, an Amazon S3 bucket) was deleted.

Permissions don't allow sending logs to the destination.

newListResolverQueryLogConfigAssociations :: ListResolverQueryLogConfigAssociations

Create a value of ListResolverQueryLogConfigAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListResolverQueryLogConfigAssociations', listResolverQueryLogConfigAssociations_filters - An optional specification to return a subset of query logging associations.

If you submit a second or subsequent ListResolverQueryLogConfigAssociations request and specify the NextToken parameter, you must use the same values for Filters, if any, as in the previous request.

$sel:nextToken:ListResolverQueryLogConfigAssociations', listResolverQueryLogConfigAssociations_nextToken - For the first ListResolverQueryLogConfigAssociations request, omit this value.

If there are more than MaxResults query logging associations that match the values that you specify for Filters, you can submit another ListResolverQueryLogConfigAssociations request to get the next group of associations. In the next request, specify the value of NextToken from the previous response.

$sel:sortOrder:ListResolverQueryLogConfigAssociations', listResolverQueryLogConfigAssociations_sortOrder - If you specified a value for SortBy, the order that you want query logging associations to be listed in, ASCENDING or DESCENDING.

If you submit a second or subsequent ListResolverQueryLogConfigAssociations request and specify the NextToken parameter, you must use the same value for SortOrder, if any, as in the previous request.

$sel:maxResults:ListResolverQueryLogConfigAssociations', listResolverQueryLogConfigAssociations_maxResults - The maximum number of query logging associations that you want to return in the response to a ListResolverQueryLogConfigAssociations request. If you don't specify a value for MaxResults, Resolver returns up to 100 query logging associations.

$sel:sortBy:ListResolverQueryLogConfigAssociations', listResolverQueryLogConfigAssociations_sortBy - The element that you want Resolver to sort query logging associations by.

If you submit a second or subsequent ListResolverQueryLogConfigAssociations request and specify the NextToken parameter, you must use the same value for SortBy, if any, as in the previous request.

Valid values include the following elements:

CreationTime: The ID of the query logging association.

Error: If the value of Status is FAILED, the value of Error indicates the cause:

DESTINATION_NOT_FOUND: The specified destination (for example, an Amazon S3 bucket) was deleted.

ACCESS_DENIED: Permissions don't allow sending logs to the destination.

If Status is a value other than FAILED, ERROR is null.

Id: The ID of the query logging association

ResolverQueryLogConfigId: The ID of the query logging configuration

ResourceId: The ID of the VPC that is associated with the query logging configuration

Status: The current status of the configuration. Valid values include the following:

CREATING: Resolver is creating an association between an Amazon VPC and a query logging configuration.

CREATED: The association between an Amazon VPC and a query logging configuration was successfully created. Resolver is logging queries that originate in the specified VPC.

DELETING: Resolver is deleting this query logging association.

FAILED: Resolver either couldn't create or couldn't delete the query logging association. Here are two common causes:

The specified destination (for example, an Amazon S3 bucket) was deleted.

Permissions don't allow sending logs to the destination.

Request Lenses

listResolverQueryLogConfigAssociations_filters :: Lens' ListResolverQueryLogConfigAssociations (Maybe [Filter])

An optional specification to return a subset of query logging associations.

If you submit a second or subsequent ListResolverQueryLogConfigAssociations request and specify the NextToken parameter, you must use the same values for Filters, if any, as in the previous request.

listResolverQueryLogConfigAssociations_nextToken :: Lens' ListResolverQueryLogConfigAssociations (Maybe Text)

For the first ListResolverQueryLogConfigAssociations request, omit this value.

If there are more than MaxResults query logging associations that match the values that you specify for Filters, you can submit another ListResolverQueryLogConfigAssociations request to get the next group of associations. In the next request, specify the value of NextToken from the previous response.

listResolverQueryLogConfigAssociations_sortOrder :: Lens' ListResolverQueryLogConfigAssociations (Maybe SortOrder)

If you specified a value for SortBy, the order that you want query logging associations to be listed in, ASCENDING or DESCENDING.

If you submit a second or subsequent ListResolverQueryLogConfigAssociations request and specify the NextToken parameter, you must use the same value for SortOrder, if any, as in the previous request.

listResolverQueryLogConfigAssociations_maxResults :: Lens' ListResolverQueryLogConfigAssociations (Maybe Natural)

The maximum number of query logging associations that you want to return in the response to a ListResolverQueryLogConfigAssociations request. If you don't specify a value for MaxResults, Resolver returns up to 100 query logging associations.

listResolverQueryLogConfigAssociations_sortBy :: Lens' ListResolverQueryLogConfigAssociations (Maybe Text)

The element that you want Resolver to sort query logging associations by.

If you submit a second or subsequent ListResolverQueryLogConfigAssociations request and specify the NextToken parameter, you must use the same value for SortBy, if any, as in the previous request.

Valid values include the following elements:

CreationTime: The ID of the query logging association.

Error: If the value of Status is FAILED, the value of Error indicates the cause:

DESTINATION_NOT_FOUND: The specified destination (for example, an Amazon S3 bucket) was deleted.

ACCESS_DENIED: Permissions don't allow sending logs to the destination.

If Status is a value other than FAILED, ERROR is null.

Id: The ID of the query logging association

ResolverQueryLogConfigId: The ID of the query logging configuration

ResourceId: The ID of the VPC that is associated with the query logging configuration

Status: The current status of the configuration. Valid values include the following:

CREATING: Resolver is creating an association between an Amazon VPC and a query logging configuration.

CREATED: The association between an Amazon VPC and a query logging configuration was successfully created. Resolver is logging queries that originate in the specified VPC.

DELETING: Resolver is deleting this query logging association.

FAILED: Resolver either couldn't create or couldn't delete the query logging association. Here are two common causes:

The specified destination (for example, an Amazon S3 bucket) was deleted.

Permissions don't allow sending logs to the destination.

Destructuring the Response

data ListResolverQueryLogConfigAssociationsResponse

See: newListResolverQueryLogConfigAssociationsResponse smart constructor.

Constructors

ListResolverQueryLogConfigAssociationsResponse'

Fields

totalFilteredCount :: Maybe Int

The total number of query logging associations that were created by the current account in the specified Region and that match the filters that were specified in the ListResolverQueryLogConfigAssociations request. For the total number of associations that were created by the current account in the specified Region, see TotalCount.

resolverQueryLogConfigAssociations :: Maybe [ResolverQueryLogConfigAssociation]

A list that contains one ResolverQueryLogConfigAssociations element for each query logging association that matches the values that you specified for Filter.

nextToken :: Maybe Text

If there are more than MaxResults query logging associations, you can submit another ListResolverQueryLogConfigAssociations request to get the next group of associations. In the next request, specify the value of NextToken from the previous response.

totalCount :: Maybe Int

The total number of query logging associations that were created by the current account in the specified Region. This count can differ from the number of associations that are returned in a ListResolverQueryLogConfigAssociations response, depending on the values that you specify in the request.

httpStatus :: Int

The response's http status code.

newListResolverQueryLogConfigAssociationsResponse

Arguments

:: Int
$sel:httpStatus:ListResolverQueryLogConfigAssociationsResponse'

-> ListResolverQueryLogConfigAssociationsResponse

Create a value of ListResolverQueryLogConfigAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:totalFilteredCount:ListResolverQueryLogConfigAssociationsResponse', listResolverQueryLogConfigAssociationsResponse_totalFilteredCount - The total number of query logging associations that were created by the current account in the specified Region and that match the filters that were specified in the ListResolverQueryLogConfigAssociations request. For the total number of associations that were created by the current account in the specified Region, see TotalCount.

$sel:resolverQueryLogConfigAssociations:ListResolverQueryLogConfigAssociationsResponse', listResolverQueryLogConfigAssociationsResponse_resolverQueryLogConfigAssociations - A list that contains one ResolverQueryLogConfigAssociations element for each query logging association that matches the values that you specified for Filter.

$sel:nextToken:ListResolverQueryLogConfigAssociations', listResolverQueryLogConfigAssociationsResponse_nextToken - If there are more than MaxResults query logging associations, you can submit another ListResolverQueryLogConfigAssociations request to get the next group of associations. In the next request, specify the value of NextToken from the previous response.

$sel:totalCount:ListResolverQueryLogConfigAssociationsResponse', listResolverQueryLogConfigAssociationsResponse_totalCount - The total number of query logging associations that were created by the current account in the specified Region. This count can differ from the number of associations that are returned in a ListResolverQueryLogConfigAssociations response, depending on the values that you specify in the request.

$sel:httpStatus:ListResolverQueryLogConfigAssociationsResponse', listResolverQueryLogConfigAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

listResolverQueryLogConfigAssociationsResponse_totalFilteredCount :: Lens' ListResolverQueryLogConfigAssociationsResponse (Maybe Int)

The total number of query logging associations that were created by the current account in the specified Region and that match the filters that were specified in the ListResolverQueryLogConfigAssociations request. For the total number of associations that were created by the current account in the specified Region, see TotalCount.

listResolverQueryLogConfigAssociationsResponse_resolverQueryLogConfigAssociations :: Lens' ListResolverQueryLogConfigAssociationsResponse (Maybe [ResolverQueryLogConfigAssociation])

A list that contains one ResolverQueryLogConfigAssociations element for each query logging association that matches the values that you specified for Filter.

listResolverQueryLogConfigAssociationsResponse_nextToken :: Lens' ListResolverQueryLogConfigAssociationsResponse (Maybe Text)

If there are more than MaxResults query logging associations, you can submit another ListResolverQueryLogConfigAssociations request to get the next group of associations. In the next request, specify the value of NextToken from the previous response.

listResolverQueryLogConfigAssociationsResponse_totalCount :: Lens' ListResolverQueryLogConfigAssociationsResponse (Maybe Int)

The total number of query logging associations that were created by the current account in the specified Region. This count can differ from the number of associations that are returned in a ListResolverQueryLogConfigAssociations response, depending on the values that you specify in the request.

listResolverQueryLogConfigAssociationsResponse_httpStatus :: Lens' ListResolverQueryLogConfigAssociationsResponse Int

The response's http status code.