Amazonka.Route53Resolver.PutResolverRulePolicy

Description

Specifies an Amazon Web Services rule that you want to share with another account, the account that you want to share the rule with, and the operations that you want the account to be able to perform on the rule.

Creating a Request

data PutResolverRulePolicy

See: newPutResolverRulePolicy smart constructor.

Constructors

PutResolverRulePolicy'

Fields

arn :: Text

The Amazon Resource Name (ARN) of the rule that you want to share with another account.

resolverRulePolicy :: Text

An Identity and Access Management policy statement that lists the rules that you want to share with another Amazon Web Services account and the operations that you want the account to be able to perform. You can specify the following operations in the Action section of the statement:

route53resolver:GetResolverRule

route53resolver:AssociateResolverRule

route53resolver:DisassociateResolverRule

route53resolver:ListResolverRules

route53resolver:ListResolverRuleAssociations

In the Resource section of the statement, specify the ARN for the rule that you want to share with another account. Specify the same ARN that you specified in Arn.

newPutResolverRulePolicy

Arguments

:: Text
$sel:arn:PutResolverRulePolicy'

-> Text
$sel:resolverRulePolicy:PutResolverRulePolicy'

-> PutResolverRulePolicy

Create a value of PutResolverRulePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:PutResolverRulePolicy', putResolverRulePolicy_arn - The Amazon Resource Name (ARN) of the rule that you want to share with another account.

$sel:resolverRulePolicy:PutResolverRulePolicy', putResolverRulePolicy_resolverRulePolicy - An Identity and Access Management policy statement that lists the rules that you want to share with another Amazon Web Services account and the operations that you want the account to be able to perform. You can specify the following operations in the Action section of the statement:

route53resolver:GetResolverRule

route53resolver:AssociateResolverRule

route53resolver:DisassociateResolverRule

route53resolver:ListResolverRules

route53resolver:ListResolverRuleAssociations

In the Resource section of the statement, specify the ARN for the rule that you want to share with another account. Specify the same ARN that you specified in Arn.

Request Lenses

putResolverRulePolicy_arn :: Lens' PutResolverRulePolicy Text

The Amazon Resource Name (ARN) of the rule that you want to share with another account.

putResolverRulePolicy_resolverRulePolicy :: Lens' PutResolverRulePolicy Text

An Identity and Access Management policy statement that lists the rules that you want to share with another Amazon Web Services account and the operations that you want the account to be able to perform. You can specify the following operations in the Action section of the statement:

route53resolver:GetResolverRule

route53resolver:AssociateResolverRule

route53resolver:DisassociateResolverRule

route53resolver:ListResolverRules

route53resolver:ListResolverRuleAssociations

In the Resource section of the statement, specify the ARN for the rule that you want to share with another account. Specify the same ARN that you specified in Arn.

Destructuring the Response

data PutResolverRulePolicyResponse

The response to a PutResolverRulePolicy request.

See: newPutResolverRulePolicyResponse smart constructor.

Constructors

PutResolverRulePolicyResponse'

Fields

returnValue :: Maybe Bool

Whether the PutResolverRulePolicy request was successful.

httpStatus :: Int

The response's http status code.

newPutResolverRulePolicyResponse

Arguments

:: Int
$sel:httpStatus:PutResolverRulePolicyResponse'

-> PutResolverRulePolicyResponse

Create a value of PutResolverRulePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:returnValue:PutResolverRulePolicyResponse', putResolverRulePolicyResponse_returnValue - Whether the PutResolverRulePolicy request was successful.

$sel:httpStatus:PutResolverRulePolicyResponse', putResolverRulePolicyResponse_httpStatus - The response's http status code.

Response Lenses

putResolverRulePolicyResponse_returnValue :: Lens' PutResolverRulePolicyResponse (Maybe Bool)

Whether the PutResolverRulePolicy request was successful.

putResolverRulePolicyResponse_httpStatus :: Lens' PutResolverRulePolicyResponse Int

The response's http status code.