Amazonka.Route53Resolver.Types.FirewallDomainList

Description

Documentation

data FirewallDomainList

High-level information about a list of firewall domains for use in a FirewallRule. This is returned by GetFirewallDomainList.

To retrieve the domains that are defined for this domain list, call ListFirewallDomains.

See: newFirewallDomainList smart constructor.

Constructors

FirewallDomainList'

Fields

creationTime :: Maybe Text

The date and time that the domain list was created, in Unix time format and Coordinated Universal Time (UTC).

status :: Maybe FirewallDomainListStatus

The status of the domain list.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall domain list.

creatorRequestId :: Maybe Text

A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of running the operation twice. This can be any unique string, for example, a timestamp.

managedOwnerName :: Maybe Text

The owner of the list, used only for lists that are not managed by you. For example, the managed domain list AWSManagedDomainsMalwareDomainList has the managed owner name Route 53 Resolver DNS Firewall.

domainCount :: Maybe Int

The number of domain names that are specified in the domain list.

modificationTime :: Maybe Text

The date and time that the domain list was last modified, in Unix time format and Coordinated Universal Time (UTC).

statusMessage :: Maybe Text

Additional information about the status of the list, if available.

name :: Maybe Text

The name of the domain list.

id :: Maybe Text

The ID of the domain list.

newFirewallDomainList :: FirewallDomainList

Create a value of FirewallDomainList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:FirewallDomainList', firewallDomainList_creationTime - The date and time that the domain list was created, in Unix time format and Coordinated Universal Time (UTC).

$sel:status:FirewallDomainList', firewallDomainList_status - The status of the domain list.

$sel:arn:FirewallDomainList', firewallDomainList_arn - The Amazon Resource Name (ARN) of the firewall domain list.

$sel:creatorRequestId:FirewallDomainList', firewallDomainList_creatorRequestId - A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of running the operation twice. This can be any unique string, for example, a timestamp.

$sel:managedOwnerName:FirewallDomainList', firewallDomainList_managedOwnerName - The owner of the list, used only for lists that are not managed by you. For example, the managed domain list AWSManagedDomainsMalwareDomainList has the managed owner name Route 53 Resolver DNS Firewall.

$sel:domainCount:FirewallDomainList', firewallDomainList_domainCount - The number of domain names that are specified in the domain list.

$sel:modificationTime:FirewallDomainList', firewallDomainList_modificationTime - The date and time that the domain list was last modified, in Unix time format and Coordinated Universal Time (UTC).

$sel:statusMessage:FirewallDomainList', firewallDomainList_statusMessage - Additional information about the status of the list, if available.

$sel:name:FirewallDomainList', firewallDomainList_name - The name of the domain list.

$sel:id:FirewallDomainList', firewallDomainList_id - The ID of the domain list.

firewallDomainList_creationTime :: Lens' FirewallDomainList (Maybe Text)

The date and time that the domain list was created, in Unix time format and Coordinated Universal Time (UTC).

firewallDomainList_status :: Lens' FirewallDomainList (Maybe FirewallDomainListStatus)

The status of the domain list.

firewallDomainList_arn :: Lens' FirewallDomainList (Maybe Text)

The Amazon Resource Name (ARN) of the firewall domain list.

firewallDomainList_creatorRequestId :: Lens' FirewallDomainList (Maybe Text)

A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of running the operation twice. This can be any unique string, for example, a timestamp.

firewallDomainList_managedOwnerName :: Lens' FirewallDomainList (Maybe Text)

The owner of the list, used only for lists that are not managed by you. For example, the managed domain list AWSManagedDomainsMalwareDomainList has the managed owner name Route 53 Resolver DNS Firewall.

firewallDomainList_domainCount :: Lens' FirewallDomainList (Maybe Int)

The number of domain names that are specified in the domain list.

firewallDomainList_modificationTime :: Lens' FirewallDomainList (Maybe Text)

The date and time that the domain list was last modified, in Unix time format and Coordinated Universal Time (UTC).

firewallDomainList_statusMessage :: Lens' FirewallDomainList (Maybe Text)

Additional information about the status of the list, if available.

firewallDomainList_name :: Lens' FirewallDomainList (Maybe Text)

The name of the domain list.

firewallDomainList_id :: Lens' FirewallDomainList (Maybe Text)

The ID of the domain list.