Amazonka.Route53Resolver.Types.FirewallRuleGroup

Description

Documentation

data FirewallRuleGroup

High-level information for a firewall rule group. A firewall rule group is a collection of rules that DNS Firewall uses to filter DNS network traffic for a VPC. To retrieve the rules for the rule group, call ListFirewallRules.

See: newFirewallRuleGroup smart constructor.

Constructors

FirewallRuleGroup'

Fields

creationTime :: Maybe Text

The date and time that the rule group was created, in Unix time format and Coordinated Universal Time (UTC).

status :: Maybe FirewallRuleGroupStatus

The status of the domain list.

arn :: Maybe Text

The ARN (Amazon Resource Name) of the rule group.

creatorRequestId :: Maybe Text

A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of running the operation twice. This can be any unique string, for example, a timestamp.

modificationTime :: Maybe Text

The date and time that the rule group was last modified, in Unix time format and Coordinated Universal Time (UTC).

shareStatus :: Maybe ShareStatus

Whether the rule group is shared with other Amazon Web Services accounts, or was shared with the current account by another Amazon Web Services account. Sharing is configured through Resource Access Manager (RAM).

ownerId :: Maybe Text

The Amazon Web Services account ID for the account that created the rule group. When a rule group is shared with your account, this is the account that has shared the rule group with you.

statusMessage :: Maybe Text

Additional information about the status of the rule group, if available.

name :: Maybe Text

The name of the rule group.

id :: Maybe Text

The ID of the rule group.

ruleCount :: Maybe Int

The number of rules in the rule group.

newFirewallRuleGroup :: FirewallRuleGroup

Create a value of FirewallRuleGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:FirewallRuleGroup', firewallRuleGroup_creationTime - The date and time that the rule group was created, in Unix time format and Coordinated Universal Time (UTC).

$sel:status:FirewallRuleGroup', firewallRuleGroup_status - The status of the domain list.

$sel:arn:FirewallRuleGroup', firewallRuleGroup_arn - The ARN (Amazon Resource Name) of the rule group.

$sel:creatorRequestId:FirewallRuleGroup', firewallRuleGroup_creatorRequestId - A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of running the operation twice. This can be any unique string, for example, a timestamp.

$sel:modificationTime:FirewallRuleGroup', firewallRuleGroup_modificationTime - The date and time that the rule group was last modified, in Unix time format and Coordinated Universal Time (UTC).

$sel:shareStatus:FirewallRuleGroup', firewallRuleGroup_shareStatus - Whether the rule group is shared with other Amazon Web Services accounts, or was shared with the current account by another Amazon Web Services account. Sharing is configured through Resource Access Manager (RAM).

$sel:ownerId:FirewallRuleGroup', firewallRuleGroup_ownerId - The Amazon Web Services account ID for the account that created the rule group. When a rule group is shared with your account, this is the account that has shared the rule group with you.

$sel:statusMessage:FirewallRuleGroup', firewallRuleGroup_statusMessage - Additional information about the status of the rule group, if available.

$sel:name:FirewallRuleGroup', firewallRuleGroup_name - The name of the rule group.

$sel:id:FirewallRuleGroup', firewallRuleGroup_id - The ID of the rule group.

$sel:ruleCount:FirewallRuleGroup', firewallRuleGroup_ruleCount - The number of rules in the rule group.

firewallRuleGroup_creationTime :: Lens' FirewallRuleGroup (Maybe Text)

The date and time that the rule group was created, in Unix time format and Coordinated Universal Time (UTC).

firewallRuleGroup_status :: Lens' FirewallRuleGroup (Maybe FirewallRuleGroupStatus)

The status of the domain list.

firewallRuleGroup_arn :: Lens' FirewallRuleGroup (Maybe Text)

The ARN (Amazon Resource Name) of the rule group.

firewallRuleGroup_creatorRequestId :: Lens' FirewallRuleGroup (Maybe Text)

A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of running the operation twice. This can be any unique string, for example, a timestamp.

firewallRuleGroup_modificationTime :: Lens' FirewallRuleGroup (Maybe Text)

The date and time that the rule group was last modified, in Unix time format and Coordinated Universal Time (UTC).

firewallRuleGroup_shareStatus :: Lens' FirewallRuleGroup (Maybe ShareStatus)

Whether the rule group is shared with other Amazon Web Services accounts, or was shared with the current account by another Amazon Web Services account. Sharing is configured through Resource Access Manager (RAM).

firewallRuleGroup_ownerId :: Lens' FirewallRuleGroup (Maybe Text)

The Amazon Web Services account ID for the account that created the rule group. When a rule group is shared with your account, this is the account that has shared the rule group with you.

firewallRuleGroup_statusMessage :: Lens' FirewallRuleGroup (Maybe Text)

Additional information about the status of the rule group, if available.

firewallRuleGroup_name :: Lens' FirewallRuleGroup (Maybe Text)

The name of the rule group.

firewallRuleGroup_id :: Lens' FirewallRuleGroup (Maybe Text)

The ID of the rule group.

firewallRuleGroup_ruleCount :: Lens' FirewallRuleGroup (Maybe Int)

The number of rules in the rule group.