Amazonka.Route53Resolver.Types.FirewallRuleGroupAssociation

Description

Documentation

data FirewallRuleGroupAssociation

An association between a firewall rule group and a VPC, which enables DNS filtering for the VPC.

See: newFirewallRuleGroupAssociation smart constructor.

Constructors

FirewallRuleGroupAssociation'

Fields

creationTime :: Maybe Text

The date and time that the association was created, in Unix time format and Coordinated Universal Time (UTC).

status :: Maybe FirewallRuleGroupAssociationStatus

The current status of the association.

mutationProtection :: Maybe MutationProtectionStatus

If enabled, this setting disallows modification or removal of the association, to help prevent against accidentally altering DNS firewall protections.

firewallRuleGroupId :: Maybe Text

The unique identifier of the firewall rule group.

priority :: Maybe Int

The setting that determines the processing order of the rule group among the rule groups that are associated with a single VPC. DNS Firewall filters VPC traffic starting from rule group with the lowest numeric priority setting.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall rule group association.

vpcId :: Maybe Text

The unique identifier of the VPC that is associated with the rule group.

creatorRequestId :: Maybe Text

A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of running the operation twice. This can be any unique string, for example, a timestamp.

managedOwnerName :: Maybe Text

The owner of the association, used only for associations that are not managed by you. If you use Firewall Manager to manage your DNS Firewalls, then this reports Firewall Manager as the managed owner.

modificationTime :: Maybe Text

The date and time that the association was last modified, in Unix time format and Coordinated Universal Time (UTC).

statusMessage :: Maybe Text

Additional information about the status of the response, if available.

name :: Maybe Text

The name of the association.

id :: Maybe Text

The identifier for the association.

newFirewallRuleGroupAssociation :: FirewallRuleGroupAssociation

Create a value of FirewallRuleGroupAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_creationTime - The date and time that the association was created, in Unix time format and Coordinated Universal Time (UTC).

$sel:status:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_status - The current status of the association.

$sel:mutationProtection:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_mutationProtection - If enabled, this setting disallows modification or removal of the association, to help prevent against accidentally altering DNS firewall protections.

$sel:firewallRuleGroupId:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_firewallRuleGroupId - The unique identifier of the firewall rule group.

$sel:priority:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_priority - The setting that determines the processing order of the rule group among the rule groups that are associated with a single VPC. DNS Firewall filters VPC traffic starting from rule group with the lowest numeric priority setting.

$sel:arn:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_arn - The Amazon Resource Name (ARN) of the firewall rule group association.

$sel:vpcId:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_vpcId - The unique identifier of the VPC that is associated with the rule group.

$sel:creatorRequestId:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_creatorRequestId - A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of running the operation twice. This can be any unique string, for example, a timestamp.

$sel:managedOwnerName:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_managedOwnerName - The owner of the association, used only for associations that are not managed by you. If you use Firewall Manager to manage your DNS Firewalls, then this reports Firewall Manager as the managed owner.

$sel:modificationTime:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_modificationTime - The date and time that the association was last modified, in Unix time format and Coordinated Universal Time (UTC).

$sel:statusMessage:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_statusMessage - Additional information about the status of the response, if available.

$sel:name:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_name - The name of the association.

$sel:id:FirewallRuleGroupAssociation', firewallRuleGroupAssociation_id - The identifier for the association.

firewallRuleGroupAssociation_creationTime :: Lens' FirewallRuleGroupAssociation (Maybe Text)

The date and time that the association was created, in Unix time format and Coordinated Universal Time (UTC).

firewallRuleGroupAssociation_status :: Lens' FirewallRuleGroupAssociation (Maybe FirewallRuleGroupAssociationStatus)

The current status of the association.

firewallRuleGroupAssociation_mutationProtection :: Lens' FirewallRuleGroupAssociation (Maybe MutationProtectionStatus)

If enabled, this setting disallows modification or removal of the association, to help prevent against accidentally altering DNS firewall protections.

firewallRuleGroupAssociation_firewallRuleGroupId :: Lens' FirewallRuleGroupAssociation (Maybe Text)

The unique identifier of the firewall rule group.

firewallRuleGroupAssociation_priority :: Lens' FirewallRuleGroupAssociation (Maybe Int)

The setting that determines the processing order of the rule group among the rule groups that are associated with a single VPC. DNS Firewall filters VPC traffic starting from rule group with the lowest numeric priority setting.

firewallRuleGroupAssociation_arn :: Lens' FirewallRuleGroupAssociation (Maybe Text)

The Amazon Resource Name (ARN) of the firewall rule group association.

firewallRuleGroupAssociation_vpcId :: Lens' FirewallRuleGroupAssociation (Maybe Text)

The unique identifier of the VPC that is associated with the rule group.

firewallRuleGroupAssociation_creatorRequestId :: Lens' FirewallRuleGroupAssociation (Maybe Text)

A unique string defined by you to identify the request. This allows you to retry failed requests without the risk of running the operation twice. This can be any unique string, for example, a timestamp.

firewallRuleGroupAssociation_managedOwnerName :: Lens' FirewallRuleGroupAssociation (Maybe Text)

The owner of the association, used only for associations that are not managed by you. If you use Firewall Manager to manage your DNS Firewalls, then this reports Firewall Manager as the managed owner.

firewallRuleGroupAssociation_modificationTime :: Lens' FirewallRuleGroupAssociation (Maybe Text)

The date and time that the association was last modified, in Unix time format and Coordinated Universal Time (UTC).

firewallRuleGroupAssociation_statusMessage :: Lens' FirewallRuleGroupAssociation (Maybe Text)

Additional information about the status of the response, if available.

firewallRuleGroupAssociation_name :: Lens' FirewallRuleGroupAssociation (Maybe Text)

The name of the association.

firewallRuleGroupAssociation_id :: Lens' FirewallRuleGroupAssociation (Maybe Text)

The identifier for the association.