Amazonka.Route53Resolver.Types.ResolverRule

Description

Documentation

data ResolverRule

For queries that originate in your VPC, detailed information about a Resolver rule, which specifies how to route DNS queries out of the VPC. The ResolverRule parameter appears in the response to a CreateResolverRule, DeleteResolverRule, GetResolverRule, ListResolverRules, or UpdateResolverRule request.

See: newResolverRule smart constructor.

Constructors

ResolverRule'

Fields

creationTime :: Maybe Text

The date and time that the Resolver rule was created, in Unix time format and Coordinated Universal Time (UTC).

status :: Maybe ResolverRuleStatus

A code that specifies the current status of the Resolver rule.

arn :: Maybe Text

The ARN (Amazon Resource Name) for the Resolver rule specified by Id.

resolverEndpointId :: Maybe Text

The ID of the endpoint that the rule is associated with.

creatorRequestId :: Maybe Text

A unique string that you specified when you created the Resolver rule. CreatorRequestId identifies the request and allows failed requests to be retried without the risk of running the operation twice.

targetIps :: Maybe (NonEmpty TargetAddress)

An array that contains the IP addresses and ports that an outbound endpoint forwards DNS queries to. Typically, these are the IP addresses of DNS resolvers on your network. Specify IPv4 addresses. IPv6 is not supported.

modificationTime :: Maybe Text

The date and time that the Resolver rule was last updated, in Unix time format and Coordinated Universal Time (UTC).

shareStatus :: Maybe ShareStatus

Whether the rule is shared and, if so, whether the current account is sharing the rule with another account, or another account is sharing the rule with the current account.

ownerId :: Maybe Text

When a rule is shared with another Amazon Web Services account, the account ID of the account that the rule is shared with.

domainName :: Maybe Text

DNS queries for this domain name are forwarded to the IP addresses that are specified in TargetIps. If a query matches multiple Resolver rules (example.com and www.example.com), the query is routed using the Resolver rule that contains the most specific domain name (www.example.com).

statusMessage :: Maybe Text

A detailed description of the status of a Resolver rule.

name :: Maybe Text

The name for the Resolver rule, which you specified when you created the Resolver rule.

id :: Maybe Text

The ID that Resolver assigned to the Resolver rule when you created it.

ruleType :: Maybe RuleTypeOption

When you want to forward DNS queries for specified domain name to resolvers on your network, specify FORWARD.

When you have a forwarding rule to forward DNS queries for a domain to your network and you want Resolver to process queries for a subdomain of that domain, specify SYSTEM.

For example, to forward DNS queries for example.com to resolvers on your network, you create a rule and specify FORWARD for RuleType. To then have Resolver process queries for apex.example.com, you create a rule and specify SYSTEM for RuleType.

Currently, only Resolver can create rules that have a value of RECURSIVE for RuleType.

newResolverRule :: ResolverRule

Create a value of ResolverRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ResolverRule', resolverRule_creationTime - The date and time that the Resolver rule was created, in Unix time format and Coordinated Universal Time (UTC).

$sel:status:ResolverRule', resolverRule_status - A code that specifies the current status of the Resolver rule.

$sel:arn:ResolverRule', resolverRule_arn - The ARN (Amazon Resource Name) for the Resolver rule specified by Id.

$sel:resolverEndpointId:ResolverRule', resolverRule_resolverEndpointId - The ID of the endpoint that the rule is associated with.

$sel:creatorRequestId:ResolverRule', resolverRule_creatorRequestId - A unique string that you specified when you created the Resolver rule. CreatorRequestId identifies the request and allows failed requests to be retried without the risk of running the operation twice.

$sel:targetIps:ResolverRule', resolverRule_targetIps - An array that contains the IP addresses and ports that an outbound endpoint forwards DNS queries to. Typically, these are the IP addresses of DNS resolvers on your network. Specify IPv4 addresses. IPv6 is not supported.

$sel:modificationTime:ResolverRule', resolverRule_modificationTime - The date and time that the Resolver rule was last updated, in Unix time format and Coordinated Universal Time (UTC).

$sel:shareStatus:ResolverRule', resolverRule_shareStatus - Whether the rule is shared and, if so, whether the current account is sharing the rule with another account, or another account is sharing the rule with the current account.

$sel:ownerId:ResolverRule', resolverRule_ownerId - When a rule is shared with another Amazon Web Services account, the account ID of the account that the rule is shared with.

$sel:domainName:ResolverRule', resolverRule_domainName - DNS queries for this domain name are forwarded to the IP addresses that are specified in TargetIps. If a query matches multiple Resolver rules (example.com and www.example.com), the query is routed using the Resolver rule that contains the most specific domain name (www.example.com).

$sel:statusMessage:ResolverRule', resolverRule_statusMessage - A detailed description of the status of a Resolver rule.

$sel:name:ResolverRule', resolverRule_name - The name for the Resolver rule, which you specified when you created the Resolver rule.

$sel:id:ResolverRule', resolverRule_id - The ID that Resolver assigned to the Resolver rule when you created it.

$sel:ruleType:ResolverRule', resolverRule_ruleType - When you want to forward DNS queries for specified domain name to resolvers on your network, specify FORWARD.

When you have a forwarding rule to forward DNS queries for a domain to your network and you want Resolver to process queries for a subdomain of that domain, specify SYSTEM.

For example, to forward DNS queries for example.com to resolvers on your network, you create a rule and specify FORWARD for RuleType. To then have Resolver process queries for apex.example.com, you create a rule and specify SYSTEM for RuleType.

Currently, only Resolver can create rules that have a value of RECURSIVE for RuleType.

resolverRule_creationTime :: Lens' ResolverRule (Maybe Text)

The date and time that the Resolver rule was created, in Unix time format and Coordinated Universal Time (UTC).

resolverRule_status :: Lens' ResolverRule (Maybe ResolverRuleStatus)

A code that specifies the current status of the Resolver rule.

resolverRule_arn :: Lens' ResolverRule (Maybe Text)

The ARN (Amazon Resource Name) for the Resolver rule specified by Id.

resolverRule_resolverEndpointId :: Lens' ResolverRule (Maybe Text)

The ID of the endpoint that the rule is associated with.

resolverRule_creatorRequestId :: Lens' ResolverRule (Maybe Text)

A unique string that you specified when you created the Resolver rule. CreatorRequestId identifies the request and allows failed requests to be retried without the risk of running the operation twice.

resolverRule_targetIps :: Lens' ResolverRule (Maybe (NonEmpty TargetAddress))

An array that contains the IP addresses and ports that an outbound endpoint forwards DNS queries to. Typically, these are the IP addresses of DNS resolvers on your network. Specify IPv4 addresses. IPv6 is not supported.

resolverRule_modificationTime :: Lens' ResolverRule (Maybe Text)

The date and time that the Resolver rule was last updated, in Unix time format and Coordinated Universal Time (UTC).

resolverRule_shareStatus :: Lens' ResolverRule (Maybe ShareStatus)

Whether the rule is shared and, if so, whether the current account is sharing the rule with another account, or another account is sharing the rule with the current account.

resolverRule_ownerId :: Lens' ResolverRule (Maybe Text)

When a rule is shared with another Amazon Web Services account, the account ID of the account that the rule is shared with.

resolverRule_domainName :: Lens' ResolverRule (Maybe Text)

DNS queries for this domain name are forwarded to the IP addresses that are specified in TargetIps. If a query matches multiple Resolver rules (example.com and www.example.com), the query is routed using the Resolver rule that contains the most specific domain name (www.example.com).

resolverRule_statusMessage :: Lens' ResolverRule (Maybe Text)

A detailed description of the status of a Resolver rule.

resolverRule_name :: Lens' ResolverRule (Maybe Text)

The name for the Resolver rule, which you specified when you created the Resolver rule.

resolverRule_id :: Lens' ResolverRule (Maybe Text)

The ID that Resolver assigned to the Resolver rule when you created it.

resolverRule_ruleType :: Lens' ResolverRule (Maybe RuleTypeOption)

When you want to forward DNS queries for specified domain name to resolvers on your network, specify FORWARD.

When you have a forwarding rule to forward DNS queries for a domain to your network and you want Resolver to process queries for a subdomain of that domain, specify SYSTEM.

For example, to forward DNS queries for example.com to resolvers on your network, you create a rule and specify FORWARD for RuleType. To then have Resolver process queries for apex.example.com, you create a rule and specify SYSTEM for RuleType.

Currently, only Resolver can create rules that have a value of RECURSIVE for RuleType.