Amazonka.Route53Resolver.Types.ResolverRuleAssociation

Description

Documentation

data ResolverRuleAssociation

In the response to an AssociateResolverRule, DisassociateResolverRule, or ListResolverRuleAssociations request, provides information about an association between a Resolver rule and a VPC. The association determines which DNS queries that originate in the VPC are forwarded to your network.

See: newResolverRuleAssociation smart constructor.

Constructors

ResolverRuleAssociation'

Fields

status :: Maybe ResolverRuleAssociationStatus

A code that specifies the current status of the association between a Resolver rule and a VPC.

resolverRuleId :: Maybe Text

The ID of the Resolver rule that you associated with the VPC that is specified by VPCId.

vPCId :: Maybe Text

The ID of the VPC that you associated the Resolver rule with.

statusMessage :: Maybe Text

A detailed description of the status of the association between a Resolver rule and a VPC.

name :: Maybe Text

The name of an association between a Resolver rule and a VPC.

id :: Maybe Text

The ID of the association between a Resolver rule and a VPC. Resolver assigns this value when you submit an AssociateResolverRule request.

newResolverRuleAssociation :: ResolverRuleAssociation

Create a value of ResolverRuleAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ResolverRuleAssociation', resolverRuleAssociation_status - A code that specifies the current status of the association between a Resolver rule and a VPC.

$sel:resolverRuleId:ResolverRuleAssociation', resolverRuleAssociation_resolverRuleId - The ID of the Resolver rule that you associated with the VPC that is specified by VPCId.

$sel:vPCId:ResolverRuleAssociation', resolverRuleAssociation_vPCId - The ID of the VPC that you associated the Resolver rule with.

$sel:statusMessage:ResolverRuleAssociation', resolverRuleAssociation_statusMessage - A detailed description of the status of the association between a Resolver rule and a VPC.

$sel:name:ResolverRuleAssociation', resolverRuleAssociation_name - The name of an association between a Resolver rule and a VPC.

$sel:id:ResolverRuleAssociation', resolverRuleAssociation_id - The ID of the association between a Resolver rule and a VPC. Resolver assigns this value when you submit an AssociateResolverRule request.

resolverRuleAssociation_status :: Lens' ResolverRuleAssociation (Maybe ResolverRuleAssociationStatus)

A code that specifies the current status of the association between a Resolver rule and a VPC.

resolverRuleAssociation_resolverRuleId :: Lens' ResolverRuleAssociation (Maybe Text)

The ID of the Resolver rule that you associated with the VPC that is specified by VPCId.

resolverRuleAssociation_vPCId :: Lens' ResolverRuleAssociation (Maybe Text)

The ID of the VPC that you associated the Resolver rule with.

resolverRuleAssociation_statusMessage :: Lens' ResolverRuleAssociation (Maybe Text)

A detailed description of the status of the association between a Resolver rule and a VPC.

resolverRuleAssociation_name :: Lens' ResolverRuleAssociation (Maybe Text)

The name of an association between a Resolver rule and a VPC.

resolverRuleAssociation_id :: Lens' ResolverRuleAssociation (Maybe Text)

The ID of the association between a Resolver rule and a VPC. Resolver assigns this value when you submit an AssociateResolverRule request.