Amazonka.Route53Resolver.UpdateFirewallRule

Description

Updates the specified firewall rule.

Creating a Request

data UpdateFirewallRule

See: newUpdateFirewallRule smart constructor.

Constructors

UpdateFirewallRule'

Fields

blockOverrideDnsType :: Maybe BlockOverrideDnsType

The DNS record's type. This determines the format of the record value that you provided in BlockOverrideDomain. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

priority :: Maybe Int

The setting that determines the processing order of the rule in the rule group. DNS Firewall processes the rules in a rule group by order of priority, starting from the lowest setting.

You must specify a unique priority for each rule in a rule group. To make it easier to insert rules later, leave space between the numbers, for example, use 100, 200, and so on. You can change the priority setting for the rules in a rule group at any time.

blockResponse :: Maybe BlockResponse

The way that you want DNS Firewall to block the request. Used for the rule action setting BLOCK.

NODATA - Respond indicating that the query was successful, but no response is available for it.

NXDOMAIN - Respond indicating that the domain name that's in the query doesn't exist.

OVERRIDE - Provide a custom override in the response. This option requires custom handling details in the rule's BlockOverride* settings.

action :: Maybe Action

The action that DNS Firewall should take on a DNS query when it matches one of the domains in the rule's domain list:

ALLOW - Permit the request to go through.

ALERT - Permit the request to go through but send an alert to the logs.

BLOCK - Disallow the request. This option requires additional details in the rule's BlockResponse.

blockOverrideTtl :: Maybe Natural

The recommended amount of time, in seconds, for the DNS resolver or web browser to cache the provided override record. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

name :: Maybe Text

The name of the rule.

blockOverrideDomain :: Maybe Text

The custom DNS record to send back in response to the query. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

firewallRuleGroupId :: Text

The unique identifier of the firewall rule group for the rule.

firewallDomainListId :: Text

The ID of the domain list to use in the rule.

newUpdateFirewallRule

Arguments

:: Text
$sel:firewallRuleGroupId:UpdateFirewallRule'

-> Text
$sel:firewallDomainListId:UpdateFirewallRule'

-> UpdateFirewallRule

Create a value of UpdateFirewallRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockOverrideDnsType:UpdateFirewallRule', updateFirewallRule_blockOverrideDnsType - The DNS record's type. This determines the format of the record value that you provided in BlockOverrideDomain. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

$sel:priority:UpdateFirewallRule', updateFirewallRule_priority - The setting that determines the processing order of the rule in the rule group. DNS Firewall processes the rules in a rule group by order of priority, starting from the lowest setting.

You must specify a unique priority for each rule in a rule group. To make it easier to insert rules later, leave space between the numbers, for example, use 100, 200, and so on. You can change the priority setting for the rules in a rule group at any time.

$sel:blockResponse:UpdateFirewallRule', updateFirewallRule_blockResponse - The way that you want DNS Firewall to block the request. Used for the rule action setting BLOCK.

NODATA - Respond indicating that the query was successful, but no response is available for it.

NXDOMAIN - Respond indicating that the domain name that's in the query doesn't exist.

OVERRIDE - Provide a custom override in the response. This option requires custom handling details in the rule's BlockOverride* settings.

$sel:action:UpdateFirewallRule', updateFirewallRule_action - The action that DNS Firewall should take on a DNS query when it matches one of the domains in the rule's domain list:

ALLOW - Permit the request to go through.

ALERT - Permit the request to go through but send an alert to the logs.

BLOCK - Disallow the request. This option requires additional details in the rule's BlockResponse.

$sel:blockOverrideTtl:UpdateFirewallRule', updateFirewallRule_blockOverrideTtl - The recommended amount of time, in seconds, for the DNS resolver or web browser to cache the provided override record. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

$sel:name:UpdateFirewallRule', updateFirewallRule_name - The name of the rule.

$sel:blockOverrideDomain:UpdateFirewallRule', updateFirewallRule_blockOverrideDomain - The custom DNS record to send back in response to the query. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

$sel:firewallRuleGroupId:UpdateFirewallRule', updateFirewallRule_firewallRuleGroupId - The unique identifier of the firewall rule group for the rule.

$sel:firewallDomainListId:UpdateFirewallRule', updateFirewallRule_firewallDomainListId - The ID of the domain list to use in the rule.

Request Lenses

updateFirewallRule_blockOverrideDnsType :: Lens' UpdateFirewallRule (Maybe BlockOverrideDnsType)

The DNS record's type. This determines the format of the record value that you provided in BlockOverrideDomain. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

updateFirewallRule_priority :: Lens' UpdateFirewallRule (Maybe Int)

The setting that determines the processing order of the rule in the rule group. DNS Firewall processes the rules in a rule group by order of priority, starting from the lowest setting.

You must specify a unique priority for each rule in a rule group. To make it easier to insert rules later, leave space between the numbers, for example, use 100, 200, and so on. You can change the priority setting for the rules in a rule group at any time.

updateFirewallRule_blockResponse :: Lens' UpdateFirewallRule (Maybe BlockResponse)

The way that you want DNS Firewall to block the request. Used for the rule action setting BLOCK.

NODATA - Respond indicating that the query was successful, but no response is available for it.

NXDOMAIN - Respond indicating that the domain name that's in the query doesn't exist.

OVERRIDE - Provide a custom override in the response. This option requires custom handling details in the rule's BlockOverride* settings.

updateFirewallRule_action :: Lens' UpdateFirewallRule (Maybe Action)

The action that DNS Firewall should take on a DNS query when it matches one of the domains in the rule's domain list:

ALLOW - Permit the request to go through.

ALERT - Permit the request to go through but send an alert to the logs.

BLOCK - Disallow the request. This option requires additional details in the rule's BlockResponse.

updateFirewallRule_blockOverrideTtl :: Lens' UpdateFirewallRule (Maybe Natural)

The recommended amount of time, in seconds, for the DNS resolver or web browser to cache the provided override record. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

updateFirewallRule_name :: Lens' UpdateFirewallRule (Maybe Text)

The name of the rule.

updateFirewallRule_blockOverrideDomain :: Lens' UpdateFirewallRule (Maybe Text)

The custom DNS record to send back in response to the query. Used for the rule action BLOCK with a BlockResponse setting of OVERRIDE.

updateFirewallRule_firewallRuleGroupId :: Lens' UpdateFirewallRule Text

The unique identifier of the firewall rule group for the rule.

updateFirewallRule_firewallDomainListId :: Lens' UpdateFirewallRule Text

The ID of the domain list to use in the rule.

Destructuring the Response

data UpdateFirewallRuleResponse

See: newUpdateFirewallRuleResponse smart constructor.

Constructors

UpdateFirewallRuleResponse'

Fields

firewallRule :: Maybe FirewallRule

The firewall rule that you just updated.

httpStatus :: Int

The response's http status code.

newUpdateFirewallRuleResponse

Arguments

:: Int
$sel:httpStatus:UpdateFirewallRuleResponse'

-> UpdateFirewallRuleResponse

Create a value of UpdateFirewallRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallRule:UpdateFirewallRuleResponse', updateFirewallRuleResponse_firewallRule - The firewall rule that you just updated.

$sel:httpStatus:UpdateFirewallRuleResponse', updateFirewallRuleResponse_httpStatus - The response's http status code.

Response Lenses

updateFirewallRuleResponse_firewallRule :: Lens' UpdateFirewallRuleResponse (Maybe FirewallRule)

The firewall rule that you just updated.

updateFirewallRuleResponse_httpStatus :: Lens' UpdateFirewallRuleResponse Int

The response's http status code.