Amazonka.Route53Resolver.UpdateResolverEndpoint

Description

Updates the name of an inbound or an outbound Resolver endpoint.

Creating a Request

data UpdateResolverEndpoint

See: newUpdateResolverEndpoint smart constructor.

Constructors

UpdateResolverEndpoint'

Fields

name :: Maybe Text

The name of the Resolver endpoint that you want to update.

resolverEndpointId :: Text

The ID of the Resolver endpoint that you want to update.

newUpdateResolverEndpoint

Arguments

:: Text
$sel:resolverEndpointId:UpdateResolverEndpoint'

-> UpdateResolverEndpoint

Create a value of UpdateResolverEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateResolverEndpoint', updateResolverEndpoint_name - The name of the Resolver endpoint that you want to update.

$sel:resolverEndpointId:UpdateResolverEndpoint', updateResolverEndpoint_resolverEndpointId - The ID of the Resolver endpoint that you want to update.

Request Lenses

updateResolverEndpoint_name :: Lens' UpdateResolverEndpoint (Maybe Text)

The name of the Resolver endpoint that you want to update.

updateResolverEndpoint_resolverEndpointId :: Lens' UpdateResolverEndpoint Text

The ID of the Resolver endpoint that you want to update.

Destructuring the Response

data UpdateResolverEndpointResponse

See: newUpdateResolverEndpointResponse smart constructor.

Constructors

UpdateResolverEndpointResponse'

Fields

resolverEndpoint :: Maybe ResolverEndpoint

The response to an UpdateResolverEndpoint request.

httpStatus :: Int

The response's http status code.

newUpdateResolverEndpointResponse

Arguments

:: Int
$sel:httpStatus:UpdateResolverEndpointResponse'

-> UpdateResolverEndpointResponse

Create a value of UpdateResolverEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolverEndpoint:UpdateResolverEndpointResponse', updateResolverEndpointResponse_resolverEndpoint - The response to an UpdateResolverEndpoint request.

$sel:httpStatus:UpdateResolverEndpointResponse', updateResolverEndpointResponse_httpStatus - The response's http status code.

Response Lenses

updateResolverEndpointResponse_resolverEndpoint :: Lens' UpdateResolverEndpointResponse (Maybe ResolverEndpoint)

The response to an UpdateResolverEndpoint request.

updateResolverEndpointResponse_httpStatus :: Lens' UpdateResolverEndpointResponse Int

The response's http status code.