Amazonka.S3.DeleteBucketCors

Description

Deletes the cors configuration information set for the bucket.

To use this operation, you must have permission to perform the s3:PutBucketCORS action. The bucket owner has this permission by default and can grant this permission to others.

For information about cors, see Enabling Cross-Origin Resource Sharing in the Amazon S3 User Guide.

Related Resources:

PutBucketCors

RESTOPTIONSobject

Creating a Request

data DeleteBucketCors

See: newDeleteBucketCors smart constructor.

Constructors

DeleteBucketCors'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

Specifies the bucket whose cors configuration is being deleted.

newDeleteBucketCors

Arguments

:: BucketName
$sel:bucket:DeleteBucketCors'

-> DeleteBucketCors

Create a value of DeleteBucketCors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:DeleteBucketCors', deleteBucketCors_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:DeleteBucketCors', deleteBucketCors_bucket - Specifies the bucket whose cors configuration is being deleted.

Request Lenses

deleteBucketCors_expectedBucketOwner :: Lens' DeleteBucketCors (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

deleteBucketCors_bucket :: Lens' DeleteBucketCors BucketName

Specifies the bucket whose cors configuration is being deleted.

Destructuring the Response

data DeleteBucketCorsResponse

See: newDeleteBucketCorsResponse smart constructor.

Constructors

DeleteBucketCorsResponse'

newDeleteBucketCorsResponse :: DeleteBucketCorsResponse

Create a value of DeleteBucketCorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.