Amazonka.S3.DeleteObjectTagging

Description

Removes the entire tag set from the specified object. For more information about managing object tags, see Object Tagging.

To use this operation, you must have permission to perform the s3:DeleteObjectTagging action.

To delete tags of a specific object version, add the versionId query parameter in the request. You will need permission for the s3:DeleteObjectVersionTagging action.

The following operations are related to DeleteBucketMetricsConfiguration:

PutObjectTagging

GetObjectTagging

Creating a Request

data DeleteObjectTagging

See: newDeleteObjectTagging smart constructor.

Constructors

DeleteObjectTagging'

Fields

versionId :: Maybe ObjectVersionId

The versionId of the object that the tag-set will be removed from.

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket name containing the objects from which to remove the tags.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

key :: ObjectKey

The key that identifies the object in the bucket from which to remove all tags.

newDeleteObjectTagging

Arguments

:: BucketName
$sel:bucket:DeleteObjectTagging'

-> ObjectKey
$sel:key:DeleteObjectTagging'

-> DeleteObjectTagging

Create a value of DeleteObjectTagging with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:DeleteObjectTagging', deleteObjectTagging_versionId - The versionId of the object that the tag-set will be removed from.

$sel:expectedBucketOwner:DeleteObjectTagging', deleteObjectTagging_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:DeleteObjectTagging', deleteObjectTagging_bucket - The bucket name containing the objects from which to remove the tags.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:key:DeleteObjectTagging', deleteObjectTagging_key - The key that identifies the object in the bucket from which to remove all tags.

Request Lenses

deleteObjectTagging_versionId :: Lens' DeleteObjectTagging (Maybe ObjectVersionId)

The versionId of the object that the tag-set will be removed from.

deleteObjectTagging_expectedBucketOwner :: Lens' DeleteObjectTagging (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

deleteObjectTagging_bucket :: Lens' DeleteObjectTagging BucketName

The bucket name containing the objects from which to remove the tags.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

deleteObjectTagging_key :: Lens' DeleteObjectTagging ObjectKey

The key that identifies the object in the bucket from which to remove all tags.

Destructuring the Response

data DeleteObjectTaggingResponse

See: newDeleteObjectTaggingResponse smart constructor.

Constructors

DeleteObjectTaggingResponse'

Fields

versionId :: Maybe ObjectVersionId

The versionId of the object the tag-set was removed from.

httpStatus :: Int

The response's http status code.

newDeleteObjectTaggingResponse

Arguments

:: Int
$sel:httpStatus:DeleteObjectTaggingResponse'

-> DeleteObjectTaggingResponse

Create a value of DeleteObjectTaggingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:DeleteObjectTagging', deleteObjectTaggingResponse_versionId - The versionId of the object the tag-set was removed from.

$sel:httpStatus:DeleteObjectTaggingResponse', deleteObjectTaggingResponse_httpStatus - The response's http status code.

Response Lenses

deleteObjectTaggingResponse_versionId :: Lens' DeleteObjectTaggingResponse (Maybe ObjectVersionId)

The versionId of the object the tag-set was removed from.

deleteObjectTaggingResponse_httpStatus :: Lens' DeleteObjectTaggingResponse Int

The response's http status code.