Amazonka.S3.GetBucketCors

Description

Returns the cors configuration information set for the bucket.

To use this operation, you must have permission to perform the s3:GetBucketCORS action. By default, the bucket owner has this permission and can grant it to others.

For more information about cors, see Enabling Cross-Origin Resource Sharing.

The following operations are related to GetBucketCors:

PutBucketCors

DeleteBucketCors

Creating a Request

data GetBucketCors

See: newGetBucketCors smart constructor.

Constructors

GetBucketCors'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket name for which to get the cors configuration.

newGetBucketCors

Arguments

:: BucketName
$sel:bucket:GetBucketCors'

-> GetBucketCors

Create a value of GetBucketCors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketCors', getBucketCors_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketCors', getBucketCors_bucket - The bucket name for which to get the cors configuration.

Request Lenses

getBucketCors_expectedBucketOwner :: Lens' GetBucketCors (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketCors_bucket :: Lens' GetBucketCors BucketName

The bucket name for which to get the cors configuration.

Destructuring the Response

data GetBucketCorsResponse

See: newGetBucketCorsResponse smart constructor.

Constructors

GetBucketCorsResponse'

Fields

cORSRules :: Maybe [CORSRule]

A set of origins and methods (cross-origin access that you want to allow). You can add up to 100 rules to the configuration.

httpStatus :: Int

The response's http status code.

newGetBucketCorsResponse

Arguments

:: Int
$sel:httpStatus:GetBucketCorsResponse'

-> GetBucketCorsResponse

Create a value of GetBucketCorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cORSRules:GetBucketCorsResponse', getBucketCorsResponse_cORSRules - A set of origins and methods (cross-origin access that you want to allow). You can add up to 100 rules to the configuration.

$sel:httpStatus:GetBucketCorsResponse', getBucketCorsResponse_httpStatus - The response's http status code.

Response Lenses

getBucketCorsResponse_cORSRules :: Lens' GetBucketCorsResponse (Maybe [CORSRule])

A set of origins and methods (cross-origin access that you want to allow). You can add up to 100 rules to the configuration.

getBucketCorsResponse_httpStatus :: Lens' GetBucketCorsResponse Int

The response's http status code.