Amazonka.S3.GetBucketIntelligentTieringConfiguration

Description

Gets the S3 Intelligent-Tiering configuration from the specified bucket.

The S3 Intelligent-Tiering storage class is designed to optimize storage costs by automatically moving data to the most cost-effective storage access tier, without performance impact or operational overhead. S3 Intelligent-Tiering delivers automatic cost savings in two low latency and high throughput access tiers. For data that can be accessed asynchronously, you can choose to activate automatic archiving capabilities within the S3 Intelligent-Tiering storage class.

The S3 Intelligent-Tiering storage class is the ideal storage class for data with unknown, changing, or unpredictable access patterns, independent of object size or retention period. If the size of an object is less than 128 KB, it is not eligible for auto-tiering. Smaller objects can be stored, but they are always charged at the Frequent Access tier rates in the S3 Intelligent-Tiering storage class.

For more information, see Storage class for automatically optimizing frequently and infrequently accessed objects.

Operations related to GetBucketIntelligentTieringConfiguration include:

DeleteBucketIntelligentTieringConfiguration

PutBucketIntelligentTieringConfiguration

ListBucketIntelligentTieringConfigurations

Creating a Request

data GetBucketIntelligentTieringConfiguration

See: newGetBucketIntelligentTieringConfiguration smart constructor.

Constructors

GetBucketIntelligentTieringConfiguration'

Fields

bucket :: BucketName

The name of the Amazon S3 bucket whose configuration you want to modify or retrieve.

id :: Text

The ID used to identify the S3 Intelligent-Tiering configuration.

newGetBucketIntelligentTieringConfiguration

Arguments

:: BucketName
$sel:bucket:GetBucketIntelligentTieringConfiguration'

-> Text
$sel:id:GetBucketIntelligentTieringConfiguration'

-> GetBucketIntelligentTieringConfiguration

Create a value of GetBucketIntelligentTieringConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucket:GetBucketIntelligentTieringConfiguration', getBucketIntelligentTieringConfiguration_bucket - The name of the Amazon S3 bucket whose configuration you want to modify or retrieve.

$sel:id:GetBucketIntelligentTieringConfiguration', getBucketIntelligentTieringConfiguration_id - The ID used to identify the S3 Intelligent-Tiering configuration.

Request Lenses

getBucketIntelligentTieringConfiguration_bucket :: Lens' GetBucketIntelligentTieringConfiguration BucketName

The name of the Amazon S3 bucket whose configuration you want to modify or retrieve.

getBucketIntelligentTieringConfiguration_id :: Lens' GetBucketIntelligentTieringConfiguration Text

The ID used to identify the S3 Intelligent-Tiering configuration.

Destructuring the Response

data GetBucketIntelligentTieringConfigurationResponse

See: newGetBucketIntelligentTieringConfigurationResponse smart constructor.

Constructors

GetBucketIntelligentTieringConfigurationResponse'

Fields

intelligentTieringConfiguration :: Maybe IntelligentTieringConfiguration

Container for S3 Intelligent-Tiering configuration.

httpStatus :: Int

The response's http status code.

newGetBucketIntelligentTieringConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetBucketIntelligentTieringConfigurationResponse'

-> GetBucketIntelligentTieringConfigurationResponse

Create a value of GetBucketIntelligentTieringConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:intelligentTieringConfiguration:GetBucketIntelligentTieringConfigurationResponse', getBucketIntelligentTieringConfigurationResponse_intelligentTieringConfiguration - Container for S3 Intelligent-Tiering configuration.

$sel:httpStatus:GetBucketIntelligentTieringConfigurationResponse', getBucketIntelligentTieringConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getBucketIntelligentTieringConfigurationResponse_intelligentTieringConfiguration :: Lens' GetBucketIntelligentTieringConfigurationResponse (Maybe IntelligentTieringConfiguration)

Container for S3 Intelligent-Tiering configuration.

getBucketIntelligentTieringConfigurationResponse_httpStatus :: Lens' GetBucketIntelligentTieringConfigurationResponse Int

The response's http status code.