Amazonka.S3.GetBucketNotificationConfiguration

Description

Returns the notification configuration of a bucket.

If notifications are not enabled on the bucket, the action returns an empty NotificationConfiguration element.

By default, you must be the bucket owner to read the notification configuration of a bucket. However, the bucket owner can use a bucket policy to grant permission to other users to read this configuration with the s3:GetBucketNotification permission.

For more information about setting and reading the notification configuration on a bucket, see Setting Up Notification of Bucket Events. For more information about bucket policies, see Using Bucket Policies.

The following action is related to GetBucketNotification:

PutBucketNotification

Creating a Request

data GetBucketNotificationConfiguration

See: newGetBucketNotificationConfiguration smart constructor.

Constructors

GetBucketNotificationConfiguration'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket for which to get the notification configuration.

newGetBucketNotificationConfiguration

Arguments

:: BucketName
$sel:bucket:GetBucketNotificationConfiguration'

-> GetBucketNotificationConfiguration

Create a value of GetBucketNotificationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketNotificationConfiguration', getBucketNotificationConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketNotificationConfiguration', getBucketNotificationConfiguration_bucket - The name of the bucket for which to get the notification configuration.

Request Lenses

getBucketNotificationConfiguration_expectedBucketOwner :: Lens' GetBucketNotificationConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketNotificationConfiguration_bucket :: Lens' GetBucketNotificationConfiguration BucketName

The name of the bucket for which to get the notification configuration.

Destructuring the Response

data NotificationConfiguration

A container for specifying the notification configuration of the bucket. If this element is empty, notifications are turned off for the bucket.

See: newNotificationConfiguration smart constructor.

Constructors

NotificationConfiguration'

Fields

queueConfigurations :: Maybe [QueueConfiguration]

The Amazon Simple Queue Service queues to publish messages to and the events for which to publish messages.

topicConfigurations :: Maybe [TopicConfiguration]

The topic to which notifications are sent and the events for which notifications are generated.

lambdaFunctionConfigurations :: Maybe [LambdaFunctionConfiguration]

Describes the Lambda functions to invoke and the events for which to invoke them.

newNotificationConfiguration :: NotificationConfiguration

Create a value of NotificationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueConfigurations:NotificationConfiguration', notificationConfiguration_queueConfigurations - The Amazon Simple Queue Service queues to publish messages to and the events for which to publish messages.

$sel:topicConfigurations:NotificationConfiguration', notificationConfiguration_topicConfigurations - The topic to which notifications are sent and the events for which notifications are generated.

$sel:lambdaFunctionConfigurations:NotificationConfiguration', notificationConfiguration_lambdaFunctionConfigurations - Describes the Lambda functions to invoke and the events for which to invoke them.

Response Lenses

notificationConfiguration_queueConfigurations :: Lens' NotificationConfiguration (Maybe [QueueConfiguration])

The Amazon Simple Queue Service queues to publish messages to and the events for which to publish messages.

notificationConfiguration_topicConfigurations :: Lens' NotificationConfiguration (Maybe [TopicConfiguration])

The topic to which notifications are sent and the events for which notifications are generated.

notificationConfiguration_lambdaFunctionConfigurations :: Lens' NotificationConfiguration (Maybe [LambdaFunctionConfiguration])

Describes the Lambda functions to invoke and the events for which to invoke them.