Amazonka.S3.GetBucketPolicyStatus

Description

Retrieves the policy status for an Amazon S3 bucket, indicating whether the bucket is public. In order to use this operation, you must have the s3:GetBucketPolicyStatus permission. For more information about Amazon S3 permissions, see Specifying Permissions in a Policy.

For more information about when Amazon S3 considers a bucket public, see The Meaning of "Public".

The following operations are related to GetBucketPolicyStatus:

Using Amazon S3 Block Public Access

GetPublicAccessBlock

PutPublicAccessBlock

DeletePublicAccessBlock

Creating a Request

data GetBucketPolicyStatus

See: newGetBucketPolicyStatus smart constructor.

Constructors

GetBucketPolicyStatus'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the Amazon S3 bucket whose policy status you want to retrieve.

newGetBucketPolicyStatus

Arguments

:: BucketName
$sel:bucket:GetBucketPolicyStatus'

-> GetBucketPolicyStatus

Create a value of GetBucketPolicyStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:GetBucketPolicyStatus', getBucketPolicyStatus_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetBucketPolicyStatus', getBucketPolicyStatus_bucket - The name of the Amazon S3 bucket whose policy status you want to retrieve.

Request Lenses

getBucketPolicyStatus_expectedBucketOwner :: Lens' GetBucketPolicyStatus (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getBucketPolicyStatus_bucket :: Lens' GetBucketPolicyStatus BucketName

The name of the Amazon S3 bucket whose policy status you want to retrieve.

Destructuring the Response

data GetBucketPolicyStatusResponse

See: newGetBucketPolicyStatusResponse smart constructor.

Constructors

GetBucketPolicyStatusResponse'

Fields

policyStatus :: Maybe PolicyStatus

The policy status for the specified bucket.

httpStatus :: Int

The response's http status code.

newGetBucketPolicyStatusResponse

Arguments

:: Int
$sel:httpStatus:GetBucketPolicyStatusResponse'

-> GetBucketPolicyStatusResponse

Create a value of GetBucketPolicyStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyStatus:GetBucketPolicyStatusResponse', getBucketPolicyStatusResponse_policyStatus - The policy status for the specified bucket.

$sel:httpStatus:GetBucketPolicyStatusResponse', getBucketPolicyStatusResponse_httpStatus - The response's http status code.

Response Lenses

getBucketPolicyStatusResponse_policyStatus :: Lens' GetBucketPolicyStatusResponse (Maybe PolicyStatus)

The policy status for the specified bucket.

getBucketPolicyStatusResponse_httpStatus :: Lens' GetBucketPolicyStatusResponse Int

The response's http status code.