Amazonka.S3.GetObjectRetention

Description

Retrieves an object's retention settings. For more information, see Locking Objects.

This action is not supported by Amazon S3 on Outposts.

Creating a Request

data GetObjectRetention

See: newGetObjectRetention smart constructor.

Constructors

GetObjectRetention'

Fields

versionId :: Maybe ObjectVersionId

The version ID for the object whose retention settings you want to retrieve.

requestPayer :: Maybe RequestPayer

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket name containing the object whose retention settings you want to retrieve.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

key :: ObjectKey

The key name for the object whose retention settings you want to retrieve.

newGetObjectRetention

Arguments

:: BucketName
$sel:bucket:GetObjectRetention'

-> ObjectKey
$sel:key:GetObjectRetention'

-> GetObjectRetention

Create a value of GetObjectRetention with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:GetObjectRetention', getObjectRetention_versionId - The version ID for the object whose retention settings you want to retrieve.

$sel:requestPayer:GetObjectRetention', getObjectRetention_requestPayer - Undocumented member.

$sel:expectedBucketOwner:GetObjectRetention', getObjectRetention_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:GetObjectRetention', getObjectRetention_bucket - The bucket name containing the object whose retention settings you want to retrieve.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

$sel:key:GetObjectRetention', getObjectRetention_key - The key name for the object whose retention settings you want to retrieve.

Request Lenses

getObjectRetention_versionId :: Lens' GetObjectRetention (Maybe ObjectVersionId)

The version ID for the object whose retention settings you want to retrieve.

getObjectRetention_requestPayer :: Lens' GetObjectRetention (Maybe RequestPayer)

Undocumented member.

getObjectRetention_expectedBucketOwner :: Lens' GetObjectRetention (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

getObjectRetention_bucket :: Lens' GetObjectRetention BucketName

The bucket name containing the object whose retention settings you want to retrieve.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

getObjectRetention_key :: Lens' GetObjectRetention ObjectKey

The key name for the object whose retention settings you want to retrieve.

Destructuring the Response

data GetObjectRetentionResponse

See: newGetObjectRetentionResponse smart constructor.

Constructors

GetObjectRetentionResponse'

Fields

retention :: Maybe ObjectLockRetention

The container element for an object's retention settings.

httpStatus :: Int

The response's http status code.

newGetObjectRetentionResponse

Arguments

:: Int
$sel:httpStatus:GetObjectRetentionResponse'

-> GetObjectRetentionResponse

Create a value of GetObjectRetentionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retention:GetObjectRetentionResponse', getObjectRetentionResponse_retention - The container element for an object's retention settings.

$sel:httpStatus:GetObjectRetentionResponse', getObjectRetentionResponse_httpStatus - The response's http status code.

Response Lenses

getObjectRetentionResponse_retention :: Lens' GetObjectRetentionResponse (Maybe ObjectLockRetention)

The container element for an object's retention settings.

getObjectRetentionResponse_httpStatus :: Lens' GetObjectRetentionResponse Int

The response's http status code.