Amazonka.S3.PutBucketAnalyticsConfiguration

Description

Sets an analytics configuration for the bucket (specified by the analytics configuration ID). You can have up to 1,000 analytics configurations per bucket.

You can choose to have storage class analysis export analysis reports sent to a comma-separated values (CSV) flat file. See the DataExport request element. Reports are updated daily and are based on the object filters that you configure. When selecting data export, you specify a destination bucket and an optional destination prefix where the file is written. You can export the data to a destination bucket in a different account. However, the destination bucket must be in the same Region as the bucket that you are making the PUT analytics configuration to. For more information, see Amazon S3 Analytics – Storage Class Analysis.

You must create a bucket policy on the destination bucket where the exported file is written to grant permissions to Amazon S3 to write objects to the bucket. For an example policy, see Granting Permissions for Amazon S3 Inventory and Storage Class Analysis.

To use this operation, you must have permissions to perform the s3:PutAnalyticsConfiguration action. The bucket owner has this permission by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

Special Errors

- HTTP Error: HTTP 400 Bad Request

Code: InvalidArgument

Cause: Invalid argument.

- HTTP Error: HTTP 400 Bad Request

Code: TooManyConfigurations

/Cause: You are attempting to create a new configuration but have already reached the 1,000-configuration limit./

- HTTP Error: HTTP 403 Forbidden

Code: AccessDenied

/Cause: You are not the owner of the specified bucket, or you do not have the s3:PutAnalyticsConfiguration bucket permission to set the configuration on the bucket./

Related Resources

GetBucketAnalyticsConfiguration

DeleteBucketAnalyticsConfiguration

ListBucketAnalyticsConfigurations

Creating a Request

data PutBucketAnalyticsConfiguration

See: newPutBucketAnalyticsConfiguration smart constructor.

Constructors

PutBucketAnalyticsConfiguration'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket to which an analytics configuration is stored.

id :: Text

The ID that identifies the analytics configuration.

analyticsConfiguration :: AnalyticsConfiguration

The configuration and any analyses for the analytics filter.

newPutBucketAnalyticsConfiguration

Arguments

:: BucketName
$sel:bucket:PutBucketAnalyticsConfiguration'

-> Text
$sel:id:PutBucketAnalyticsConfiguration'

-> AnalyticsConfiguration
$sel:analyticsConfiguration:PutBucketAnalyticsConfiguration'

-> PutBucketAnalyticsConfiguration

Create a value of PutBucketAnalyticsConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:PutBucketAnalyticsConfiguration', putBucketAnalyticsConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:PutBucketAnalyticsConfiguration', putBucketAnalyticsConfiguration_bucket - The name of the bucket to which an analytics configuration is stored.

$sel:id:PutBucketAnalyticsConfiguration', putBucketAnalyticsConfiguration_id - The ID that identifies the analytics configuration.

$sel:analyticsConfiguration:PutBucketAnalyticsConfiguration', putBucketAnalyticsConfiguration_analyticsConfiguration - The configuration and any analyses for the analytics filter.

Request Lenses

putBucketAnalyticsConfiguration_expectedBucketOwner :: Lens' PutBucketAnalyticsConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketAnalyticsConfiguration_bucket :: Lens' PutBucketAnalyticsConfiguration BucketName

The name of the bucket to which an analytics configuration is stored.

putBucketAnalyticsConfiguration_id :: Lens' PutBucketAnalyticsConfiguration Text

The ID that identifies the analytics configuration.

putBucketAnalyticsConfiguration_analyticsConfiguration :: Lens' PutBucketAnalyticsConfiguration AnalyticsConfiguration

The configuration and any analyses for the analytics filter.

Destructuring the Response

data PutBucketAnalyticsConfigurationResponse

See: newPutBucketAnalyticsConfigurationResponse smart constructor.

Constructors

PutBucketAnalyticsConfigurationResponse'

newPutBucketAnalyticsConfigurationResponse :: PutBucketAnalyticsConfigurationResponse

Create a value of PutBucketAnalyticsConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.