Amazonka.S3.PutBucketIntelligentTieringConfiguration

Description

Puts a S3 Intelligent-Tiering configuration to the specified bucket. You can have up to 1,000 S3 Intelligent-Tiering configurations per bucket.

The S3 Intelligent-Tiering storage class is designed to optimize storage costs by automatically moving data to the most cost-effective storage access tier, without performance impact or operational overhead. S3 Intelligent-Tiering delivers automatic cost savings in two low latency and high throughput access tiers. For data that can be accessed asynchronously, you can choose to activate automatic archiving capabilities within the S3 Intelligent-Tiering storage class.

The S3 Intelligent-Tiering storage class is the ideal storage class for data with unknown, changing, or unpredictable access patterns, independent of object size or retention period. If the size of an object is less than 128 KB, it is not eligible for auto-tiering. Smaller objects can be stored, but they are always charged at the Frequent Access tier rates in the S3 Intelligent-Tiering storage class.

For more information, see Storage class for automatically optimizing frequently and infrequently accessed objects.

Operations related to PutBucketIntelligentTieringConfiguration include:

DeleteBucketIntelligentTieringConfiguration

GetBucketIntelligentTieringConfiguration

ListBucketIntelligentTieringConfigurations

You only need S3 Intelligent-Tiering enabled on a bucket if you want to automatically move objects stored in the S3 Intelligent-Tiering storage class to the Archive Access or Deep Archive Access tier.

Special Errors

HTTP 400 Bad Request Error

Code: InvalidArgument

Cause: Invalid Argument

HTTP 400 Bad Request Error

Code: TooManyConfigurations

Cause: You are attempting to create a new configuration but have already reached the 1,000-configuration limit.

HTTP 403 Forbidden Error

Code: AccessDenied

Cause: You are not the owner of the specified bucket, or you do not have the s3:PutIntelligentTieringConfiguration bucket permission to set the configuration on the bucket.

Creating a Request

data PutBucketIntelligentTieringConfiguration

See: newPutBucketIntelligentTieringConfiguration smart constructor.

Constructors

PutBucketIntelligentTieringConfiguration'

Fields

bucket :: BucketName

The name of the Amazon S3 bucket whose configuration you want to modify or retrieve.

id :: Text

The ID used to identify the S3 Intelligent-Tiering configuration.

intelligentTieringConfiguration :: IntelligentTieringConfiguration

Container for S3 Intelligent-Tiering configuration.

newPutBucketIntelligentTieringConfiguration

Arguments

:: BucketName
$sel:bucket:PutBucketIntelligentTieringConfiguration'

-> Text
$sel:id:PutBucketIntelligentTieringConfiguration'

-> IntelligentTieringConfiguration
$sel:intelligentTieringConfiguration:PutBucketIntelligentTieringConfiguration'

-> PutBucketIntelligentTieringConfiguration

Create a value of PutBucketIntelligentTieringConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucket:PutBucketIntelligentTieringConfiguration', putBucketIntelligentTieringConfiguration_bucket - The name of the Amazon S3 bucket whose configuration you want to modify or retrieve.

$sel:id:PutBucketIntelligentTieringConfiguration', putBucketIntelligentTieringConfiguration_id - The ID used to identify the S3 Intelligent-Tiering configuration.

$sel:intelligentTieringConfiguration:PutBucketIntelligentTieringConfiguration', putBucketIntelligentTieringConfiguration_intelligentTieringConfiguration - Container for S3 Intelligent-Tiering configuration.

Request Lenses

putBucketIntelligentTieringConfiguration_bucket :: Lens' PutBucketIntelligentTieringConfiguration BucketName

The name of the Amazon S3 bucket whose configuration you want to modify or retrieve.

putBucketIntelligentTieringConfiguration_id :: Lens' PutBucketIntelligentTieringConfiguration Text

The ID used to identify the S3 Intelligent-Tiering configuration.

putBucketIntelligentTieringConfiguration_intelligentTieringConfiguration :: Lens' PutBucketIntelligentTieringConfiguration IntelligentTieringConfiguration

Container for S3 Intelligent-Tiering configuration.

Destructuring the Response

data PutBucketIntelligentTieringConfigurationResponse

See: newPutBucketIntelligentTieringConfigurationResponse smart constructor.

Constructors

PutBucketIntelligentTieringConfigurationResponse'

newPutBucketIntelligentTieringConfigurationResponse :: PutBucketIntelligentTieringConfigurationResponse

Create a value of PutBucketIntelligentTieringConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.