Amazonka.S3.PutBucketLifecycleConfiguration

Description

Creates a new lifecycle configuration for the bucket or replaces an existing lifecycle configuration. For information about lifecycle configuration, see Managing your storage lifecycle.

Bucket lifecycle configuration now supports specifying a lifecycle rule using an object key name prefix, one or more object tags, or a combination of both. Accordingly, this section describes the latest API. The previous version of the API supported filtering based only on an object key name prefix, which is supported for backward compatibility. For the related API description, see PutBucketLifecycle.

Rules

You specify the lifecycle configuration in your request body. The lifecycle configuration is specified as XML consisting of one or more rules. Each rule consists of the following:

Filter identifying a subset of objects to which the rule applies. The filter can be based on a key name prefix, object tags, or a combination of both.

Status whether the rule is in effect.

One or more lifecycle transition and expiration actions that you want Amazon S3 to perform on the objects identified by the filter. If the state of your bucket is versioning-enabled or versioning-suspended, you can have many versions of the same object (one current version and zero or more noncurrent versions). Amazon S3 provides predefined actions that you can specify for current and noncurrent object versions.

For more information, see Object Lifecycle Management and Lifecycle Configuration Elements.

Permissions

By default, all Amazon S3 resources are private, including buckets, objects, and related subresources (for example, lifecycle configuration and website configuration). Only the resource owner (that is, the Amazon Web Services account that created it) can access the resource. The resource owner can optionally grant access permissions to others by writing an access policy. For this operation, a user must get the s3:PutLifecycleConfiguration permission.

You can also explicitly deny permissions. Explicit deny also supersedes any other permissions. If you want to block users or accounts from removing or deleting objects from your bucket, you must deny them permissions for the following actions:

s3:DeleteObject

s3:DeleteObjectVersion

s3:PutLifecycleConfiguration

For more information about permissions, see Managing Access Permissions to Your Amazon S3 Resources.

The following are related to PutBucketLifecycleConfiguration:

Examples of Lifecycle Configuration

GetBucketLifecycleConfiguration

DeleteBucketLifecycle

Creating a Request

data PutBucketLifecycleConfiguration

See: newPutBucketLifecycleConfiguration smart constructor.

Constructors

PutBucketLifecycleConfiguration'

Fields

lifecycleConfiguration :: Maybe BucketLifecycleConfiguration

Container for lifecycle rules. You can add as many as 1,000 rules.

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket for which to set the configuration.

newPutBucketLifecycleConfiguration

Arguments

:: BucketName
$sel:bucket:PutBucketLifecycleConfiguration'

-> PutBucketLifecycleConfiguration

Create a value of PutBucketLifecycleConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lifecycleConfiguration:PutBucketLifecycleConfiguration', putBucketLifecycleConfiguration_lifecycleConfiguration - Container for lifecycle rules. You can add as many as 1,000 rules.

$sel:expectedBucketOwner:PutBucketLifecycleConfiguration', putBucketLifecycleConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:PutBucketLifecycleConfiguration', putBucketLifecycleConfiguration_bucket - The name of the bucket for which to set the configuration.

Request Lenses

putBucketLifecycleConfiguration_lifecycleConfiguration :: Lens' PutBucketLifecycleConfiguration (Maybe BucketLifecycleConfiguration)

Container for lifecycle rules. You can add as many as 1,000 rules.

putBucketLifecycleConfiguration_expectedBucketOwner :: Lens' PutBucketLifecycleConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketLifecycleConfiguration_bucket :: Lens' PutBucketLifecycleConfiguration BucketName

The name of the bucket for which to set the configuration.

Destructuring the Response

data PutBucketLifecycleConfigurationResponse

See: newPutBucketLifecycleConfigurationResponse smart constructor.

Constructors

PutBucketLifecycleConfigurationResponse'

newPutBucketLifecycleConfigurationResponse :: PutBucketLifecycleConfigurationResponse

Create a value of PutBucketLifecycleConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.