Amazonka.S3.PutBucketMetricsConfiguration

Description

Sets a metrics configuration (specified by the metrics configuration ID) for the bucket. You can have up to 1,000 metrics configurations per bucket. If you're updating an existing metrics configuration, note that this is a full replacement of the existing metrics configuration. If you don't include the elements you want to keep, they are erased.

To use this operation, you must have permissions to perform the s3:PutMetricsConfiguration action. The bucket owner has this permission by default. The bucket owner can grant this permission to others. For more information about permissions, see Permissions Related to Bucket Subresource Operations and Managing Access Permissions to Your Amazon S3 Resources.

For information about CloudWatch request metrics for Amazon S3, see Monitoring Metrics with Amazon CloudWatch.

The following operations are related to PutBucketMetricsConfiguration:

DeleteBucketMetricsConfiguration

GetBucketMetricsConfiguration

ListBucketMetricsConfigurations

GetBucketLifecycle has the following special error:

Error code: TooManyConfigurations

Description: You are attempting to create a new configuration but have already reached the 1,000-configuration limit.

HTTP Status Code: HTTP 400 Bad Request

Creating a Request

data PutBucketMetricsConfiguration

See: newPutBucketMetricsConfiguration smart constructor.

Constructors

PutBucketMetricsConfiguration'

Fields

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket for which the metrics configuration is set.

id :: Text

The ID used to identify the metrics configuration.

metricsConfiguration :: MetricsConfiguration

Specifies the metrics configuration.

newPutBucketMetricsConfiguration

Arguments

:: BucketName
$sel:bucket:PutBucketMetricsConfiguration'

-> Text
$sel:id:PutBucketMetricsConfiguration'

-> MetricsConfiguration
$sel:metricsConfiguration:PutBucketMetricsConfiguration'

-> PutBucketMetricsConfiguration

Create a value of PutBucketMetricsConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedBucketOwner:PutBucketMetricsConfiguration', putBucketMetricsConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:PutBucketMetricsConfiguration', putBucketMetricsConfiguration_bucket - The name of the bucket for which the metrics configuration is set.

$sel:id:PutBucketMetricsConfiguration', putBucketMetricsConfiguration_id - The ID used to identify the metrics configuration.

$sel:metricsConfiguration:PutBucketMetricsConfiguration', putBucketMetricsConfiguration_metricsConfiguration - Specifies the metrics configuration.

Request Lenses

putBucketMetricsConfiguration_expectedBucketOwner :: Lens' PutBucketMetricsConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketMetricsConfiguration_bucket :: Lens' PutBucketMetricsConfiguration BucketName

The name of the bucket for which the metrics configuration is set.

putBucketMetricsConfiguration_id :: Lens' PutBucketMetricsConfiguration Text

The ID used to identify the metrics configuration.

putBucketMetricsConfiguration_metricsConfiguration :: Lens' PutBucketMetricsConfiguration MetricsConfiguration

Specifies the metrics configuration.

Destructuring the Response

data PutBucketMetricsConfigurationResponse

See: newPutBucketMetricsConfigurationResponse smart constructor.

Constructors

PutBucketMetricsConfigurationResponse'

newPutBucketMetricsConfigurationResponse :: PutBucketMetricsConfigurationResponse

Create a value of PutBucketMetricsConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.