Amazonka.S3.PutBucketOwnershipControls

Description

Creates or modifies OwnershipControls for an Amazon S3 bucket. To use this operation, you must have the s3:PutBucketOwnershipControls permission. For more information about Amazon S3 permissions, see Specifying Permissions in a Policy.

For information about Amazon S3 Object Ownership, see Using Object Ownership.

The following operations are related to PutBucketOwnershipControls:

GetBucketOwnershipControls

DeleteBucketOwnershipControls

Creating a Request

data PutBucketOwnershipControls

See: newPutBucketOwnershipControls smart constructor.

Constructors

PutBucketOwnershipControls'

Fields

contentMD5 :: Maybe Text

The MD5 hash of the OwnershipControls request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the Amazon S3 bucket whose OwnershipControls you want to set.

ownershipControls :: OwnershipControls

The OwnershipControls (BucketOwnerPreferred or ObjectWriter) that you want to apply to this Amazon S3 bucket.

newPutBucketOwnershipControls

Arguments

:: BucketName
$sel:bucket:PutBucketOwnershipControls'

-> OwnershipControls
$sel:ownershipControls:PutBucketOwnershipControls'

-> PutBucketOwnershipControls

Create a value of PutBucketOwnershipControls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentMD5:PutBucketOwnershipControls', putBucketOwnershipControls_contentMD5 - The MD5 hash of the OwnershipControls request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

$sel:expectedBucketOwner:PutBucketOwnershipControls', putBucketOwnershipControls_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:PutBucketOwnershipControls', putBucketOwnershipControls_bucket - The name of the Amazon S3 bucket whose OwnershipControls you want to set.

$sel:ownershipControls:PutBucketOwnershipControls', putBucketOwnershipControls_ownershipControls - The OwnershipControls (BucketOwnerPreferred or ObjectWriter) that you want to apply to this Amazon S3 bucket.

Request Lenses

putBucketOwnershipControls_contentMD5 :: Lens' PutBucketOwnershipControls (Maybe Text)

The MD5 hash of the OwnershipControls request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

putBucketOwnershipControls_expectedBucketOwner :: Lens' PutBucketOwnershipControls (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putBucketOwnershipControls_bucket :: Lens' PutBucketOwnershipControls BucketName

The name of the Amazon S3 bucket whose OwnershipControls you want to set.

putBucketOwnershipControls_ownershipControls :: Lens' PutBucketOwnershipControls OwnershipControls

The OwnershipControls (BucketOwnerPreferred or ObjectWriter) that you want to apply to this Amazon S3 bucket.

Destructuring the Response

data PutBucketOwnershipControlsResponse

See: newPutBucketOwnershipControlsResponse smart constructor.

Constructors

PutBucketOwnershipControlsResponse'

newPutBucketOwnershipControlsResponse :: PutBucketOwnershipControlsResponse

Create a value of PutBucketOwnershipControlsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.