Amazonka.S3.PutObject

Description

Adds an object to a bucket. You must have WRITE permissions on a bucket to add an object to it.

Amazon S3 never adds partial objects; if you receive a success response, Amazon S3 added the entire object to the bucket.

Amazon S3 is a distributed system. If it receives multiple write requests for the same object simultaneously, it overwrites all but the last object written. Amazon S3 does not provide object locking; if you need this, make sure to build it into your application layer or use versioning instead.

To ensure that data is not corrupted traversing the network, use the Content-MD5 header. When you use this header, Amazon S3 checks the object against the provided MD5 value and, if they do not match, returns an error. Additionally, you can calculate the MD5 while putting an object to Amazon S3 and compare the returned ETag to the calculated MD5 value.

To successfully complete the PutObject request, you must have the s3:PutObject in your IAM permissions.

To successfully change the objects acl of your PutObject request, you must have the s3:PutObjectAcl in your IAM permissions.

The Content-MD5 header is required for any request to upload an object with a retention period configured using Amazon S3 Object Lock. For more information about Amazon S3 Object Lock, see Amazon S3 Object Lock Overview in the Amazon S3 User Guide.

Server-side Encryption

You can optionally request server-side encryption. With server-side encryption, Amazon S3 encrypts your data as it writes it to disks in its data centers and decrypts the data when you access it. You have the option to provide your own encryption key or use Amazon Web Services managed encryption keys (SSE-S3 or SSE-KMS). For more information, see Using Server-Side Encryption.

If you request server-side encryption using Amazon Web Services Key Management Service (SSE-KMS), you can enable an S3 Bucket Key at the object-level. For more information, see Amazon S3 Bucket Keys in the Amazon S3 User Guide.

Access Control List (ACL)-Specific Request Headers

You can use headers to grant ACL- based permissions. By default, all objects are private. Only the owner has full access control. When adding a new object, you can grant permissions to individual Amazon Web Services accounts or to predefined groups defined by Amazon S3. These permissions are then added to the ACL on the object. For more information, see Access Control List (ACL) Overview and Managing ACLs Using the REST API.

Storage Class Options

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

Versioning

If you enable versioning for a bucket, Amazon S3 automatically generates a unique version ID for the object being stored. Amazon S3 returns this ID in the response. When you enable versioning for a bucket, if Amazon S3 receives multiple write requests for the same object simultaneously, it stores all of the objects.

For more information about versioning, see Adding Objects to Versioning Enabled Buckets. For information about returning the versioning state of a bucket, see GetBucketVersioning.

Related Resources

CopyObject

DeleteObject

Creating a Request

data PutObject

See: newPutObject smart constructor.

Constructors

PutObject'

Fields

contentLength :: Maybe Integer

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.13.

objectLockMode :: Maybe ObjectLockMode

The Object Lock mode that you want to apply to this object.

expires :: Maybe ISO8601

The date and time at which the object is no longer cacheable. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.21.

grantReadACP :: Maybe Text

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

sSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use to when encrypting the object (for example, AES256).

sSECustomerKey :: Maybe (Sensitive Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

requestPayer :: Maybe RequestPayer

grantWriteACP :: Maybe Text

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

bucketKeyEnabled :: Maybe Bool

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with a PUT action doesn’t affect bucket-level settings for S3 Bucket Key.

websiteRedirectLocation :: Maybe Text

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata. For information about object metadata, see Object Key and Metadata.

In the following example, the request header sets the redirect to an object (anotherPage.html) in the same bucket:

x-amz-website-redirect-location: /anotherPage.html

In the following example, the request header sets the object redirect to another website:

x-amz-website-redirect-location: http://www.example.com/

For more information about website hosting in Amazon S3, see Hosting Websites on Amazon S3 and How to Configure Website Page Redirects.

grantRead :: Maybe Text

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

storageClass :: Maybe StorageClass

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

sSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

sSEKMSKeyId :: Maybe (Sensitive Text)

If x-amz-server-side-encryption is present and has the value of aws:kms, this header specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetrical customer managed key that was used for the object. If you specify x-amz-server-side-encryption:aws:kms, but do not provide x-amz-server-side-encryption-aws-kms-key-id, Amazon S3 uses the Amazon Web Services managed key to protect the data. If the KMS key does not exist in the same account issuing the command, you must use the full ARN and not just the ID.

grantFullControl :: Maybe Text

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

contentEncoding :: Maybe Text

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.11.

tagging :: Maybe Text

The tag-set for the object. The tag-set must be encoded as URL Query parameters. (For example, "Key1=Value1")

contentMD5 :: Maybe Text

The base64-encoded 128-bit MD5 digest of the message (without the headers) according to RFC 1864. This header can be used as a message integrity check to verify that the data is the same data that was originally sent. Although it is optional, we recommend using the Content-MD5 mechanism as an end-to-end integrity check. For more information about REST request authentication, see REST Authentication.

objectLockRetainUntilDate :: Maybe ISO8601

The date and time when you want this object's Object Lock to expire. Must be formatted as a timestamp parameter.

metadata :: HashMap Text Text

A map of metadata to store with the object in S3.

sSEKMSEncryptionContext :: Maybe (Sensitive Text)

Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

cacheControl :: Maybe Text

Can be used to specify caching behavior along the request/reply chain. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

contentLanguage :: Maybe Text

The language the content is in.

objectLockLegalHoldStatus :: Maybe ObjectLockLegalHoldStatus

Specifies whether a legal hold will be applied to this object. For more information about S3 Object Lock, see Object Lock.

acl :: Maybe ObjectCannedACL

The canned ACL to apply to the object. For more information, see Canned ACL.

This action is not supported by Amazon S3 on Outposts.

contentDisposition :: Maybe Text

Specifies presentational information for the object. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec19.html#sec19.5.1.

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

serverSideEncryption :: Maybe ServerSideEncryption

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

contentType :: Maybe Text

A standard MIME type describing the format of the contents. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.17.

bucket :: BucketName

The bucket name to which the PUT action was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

key :: ObjectKey

Object key for which the PUT action was initiated.

body :: RequestBody

Object data.

newPutObject

Arguments

:: BucketName
$sel:bucket:PutObject'

-> ObjectKey
$sel:key:PutObject'

-> RequestBody
$sel:body:PutObject'

-> PutObject

Create a value of PutObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentLength:PutObject', putObject_contentLength - Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.13.

$sel:objectLockMode:PutObject', putObject_objectLockMode - The Object Lock mode that you want to apply to this object.

$sel:expires:PutObject', putObject_expires - The date and time at which the object is no longer cacheable. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.21.

$sel:grantReadACP:PutObject', putObject_grantReadACP - Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

$sel:sSECustomerAlgorithm:PutObject', putObject_sSECustomerAlgorithm - Specifies the algorithm to use to when encrypting the object (for example, AES256).

$sel:sSECustomerKey:PutObject', putObject_sSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

$sel:requestPayer:PutObject', putObject_requestPayer - Undocumented member.

$sel:grantWriteACP:PutObject', putObject_grantWriteACP - Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

$sel:bucketKeyEnabled:PutObject', putObject_bucketKeyEnabled - Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with a PUT action doesn’t affect bucket-level settings for S3 Bucket Key.

$sel:websiteRedirectLocation:PutObject', putObject_websiteRedirectLocation - If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata. For information about object metadata, see Object Key and Metadata.

In the following example, the request header sets the redirect to an object (anotherPage.html) in the same bucket:

x-amz-website-redirect-location: /anotherPage.html

In the following example, the request header sets the object redirect to another website:

x-amz-website-redirect-location: http://www.example.com/

For more information about website hosting in Amazon S3, see Hosting Websites on Amazon S3 and How to Configure Website Page Redirects.

$sel:grantRead:PutObject', putObject_grantRead - Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

$sel:storageClass:PutObject', putObject_storageClass - By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

$sel:sSECustomerKeyMD5:PutObject', putObject_sSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

$sel:sSEKMSKeyId:PutObject', putObject_sSEKMSKeyId - If x-amz-server-side-encryption is present and has the value of aws:kms, this header specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetrical customer managed key that was used for the object. If you specify x-amz-server-side-encryption:aws:kms, but do not provide x-amz-server-side-encryption-aws-kms-key-id, Amazon S3 uses the Amazon Web Services managed key to protect the data. If the KMS key does not exist in the same account issuing the command, you must use the full ARN and not just the ID.

$sel:grantFullControl:PutObject', putObject_grantFullControl - Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

$sel:contentEncoding:PutObject', putObject_contentEncoding - Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.11.

$sel:tagging:PutObject', putObject_tagging - The tag-set for the object. The tag-set must be encoded as URL Query parameters. (For example, "Key1=Value1")

$sel:contentMD5:PutObject', putObject_contentMD5 - The base64-encoded 128-bit MD5 digest of the message (without the headers) according to RFC 1864. This header can be used as a message integrity check to verify that the data is the same data that was originally sent. Although it is optional, we recommend using the Content-MD5 mechanism as an end-to-end integrity check. For more information about REST request authentication, see REST Authentication.

$sel:objectLockRetainUntilDate:PutObject', putObject_objectLockRetainUntilDate - The date and time when you want this object's Object Lock to expire. Must be formatted as a timestamp parameter.

$sel:metadata:PutObject', putObject_metadata - A map of metadata to store with the object in S3.

$sel:sSEKMSEncryptionContext:PutObject', putObject_sSEKMSEncryptionContext - Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

$sel:cacheControl:PutObject', putObject_cacheControl - Can be used to specify caching behavior along the request/reply chain. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

$sel:contentLanguage:PutObject', putObject_contentLanguage - The language the content is in.

$sel:objectLockLegalHoldStatus:PutObject', putObject_objectLockLegalHoldStatus - Specifies whether a legal hold will be applied to this object. For more information about S3 Object Lock, see Object Lock.

$sel:acl:PutObject', putObject_acl - The canned ACL to apply to the object. For more information, see Canned ACL.

This action is not supported by Amazon S3 on Outposts.

$sel:contentDisposition:PutObject', putObject_contentDisposition - Specifies presentational information for the object. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec19.html#sec19.5.1.

$sel:expectedBucketOwner:PutObject', putObject_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:serverSideEncryption:PutObject', putObject_serverSideEncryption - The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

$sel:contentType:PutObject', putObject_contentType - A standard MIME type describing the format of the contents. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.17.

$sel:bucket:PutObject', putObject_bucket - The bucket name to which the PUT action was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:key:PutObject', putObject_key - Object key for which the PUT action was initiated.

$sel:body:PutObject', putObject_body - Object data.

Request Lenses

putObject_contentLength :: Lens' PutObject (Maybe Integer)

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.13.

putObject_objectLockMode :: Lens' PutObject (Maybe ObjectLockMode)

The Object Lock mode that you want to apply to this object.

putObject_expires :: Lens' PutObject (Maybe UTCTime)

The date and time at which the object is no longer cacheable. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.21.

putObject_grantReadACP :: Lens' PutObject (Maybe Text)

Allows grantee to read the object ACL.

This action is not supported by Amazon S3 on Outposts.

putObject_sSECustomerAlgorithm :: Lens' PutObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (for example, AES256).

putObject_sSECustomerKey :: Lens' PutObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header.

putObject_requestPayer :: Lens' PutObject (Maybe RequestPayer)

Undocumented member.

putObject_grantWriteACP :: Lens' PutObject (Maybe Text)

Allows grantee to write the ACL for the applicable object.

This action is not supported by Amazon S3 on Outposts.

putObject_bucketKeyEnabled :: Lens' PutObject (Maybe Bool)

Specifies whether Amazon S3 should use an S3 Bucket Key for object encryption with server-side encryption using AWS KMS (SSE-KMS). Setting this header to true causes Amazon S3 to use an S3 Bucket Key for object encryption with SSE-KMS.

Specifying this header with a PUT action doesn’t affect bucket-level settings for S3 Bucket Key.

putObject_websiteRedirectLocation :: Lens' PutObject (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata. For information about object metadata, see Object Key and Metadata.

In the following example, the request header sets the redirect to an object (anotherPage.html) in the same bucket:

x-amz-website-redirect-location: /anotherPage.html

In the following example, the request header sets the object redirect to another website:

x-amz-website-redirect-location: http://www.example.com/

For more information about website hosting in Amazon S3, see Hosting Websites on Amazon S3 and How to Configure Website Page Redirects.

putObject_grantRead :: Lens' PutObject (Maybe Text)

Allows grantee to read the object data and its metadata.

This action is not supported by Amazon S3 on Outposts.

putObject_storageClass :: Lens' PutObject (Maybe StorageClass)

By default, Amazon S3 uses the STANDARD Storage Class to store newly created objects. The STANDARD storage class provides high durability and high availability. Depending on performance needs, you can specify a different Storage Class. Amazon S3 on Outposts only uses the OUTPOSTS Storage Class. For more information, see Storage Classes in the Amazon S3 User Guide.

putObject_sSECustomerKeyMD5 :: Lens' PutObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

putObject_sSEKMSKeyId :: Lens' PutObject (Maybe Text)

If x-amz-server-side-encryption is present and has the value of aws:kms, this header specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetrical customer managed key that was used for the object. If you specify x-amz-server-side-encryption:aws:kms, but do not provide x-amz-server-side-encryption-aws-kms-key-id, Amazon S3 uses the Amazon Web Services managed key to protect the data. If the KMS key does not exist in the same account issuing the command, you must use the full ARN and not just the ID.

putObject_grantFullControl :: Lens' PutObject (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

This action is not supported by Amazon S3 on Outposts.

putObject_contentEncoding :: Lens' PutObject (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.11.

putObject_tagging :: Lens' PutObject (Maybe Text)

The tag-set for the object. The tag-set must be encoded as URL Query parameters. (For example, "Key1=Value1")

putObject_contentMD5 :: Lens' PutObject (Maybe Text)

The base64-encoded 128-bit MD5 digest of the message (without the headers) according to RFC 1864. This header can be used as a message integrity check to verify that the data is the same data that was originally sent. Although it is optional, we recommend using the Content-MD5 mechanism as an end-to-end integrity check. For more information about REST request authentication, see REST Authentication.

putObject_objectLockRetainUntilDate :: Lens' PutObject (Maybe UTCTime)

The date and time when you want this object's Object Lock to expire. Must be formatted as a timestamp parameter.

putObject_metadata :: Lens' PutObject (HashMap Text Text)

A map of metadata to store with the object in S3.

putObject_sSEKMSEncryptionContext :: Lens' PutObject (Maybe Text)

Specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

putObject_cacheControl :: Lens' PutObject (Maybe Text)

Can be used to specify caching behavior along the request/reply chain. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

putObject_contentLanguage :: Lens' PutObject (Maybe Text)

The language the content is in.

putObject_objectLockLegalHoldStatus :: Lens' PutObject (Maybe ObjectLockLegalHoldStatus)

Specifies whether a legal hold will be applied to this object. For more information about S3 Object Lock, see Object Lock.

putObject_acl :: Lens' PutObject (Maybe ObjectCannedACL)

The canned ACL to apply to the object. For more information, see Canned ACL.

This action is not supported by Amazon S3 on Outposts.

putObject_contentDisposition :: Lens' PutObject (Maybe Text)

Specifies presentational information for the object. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec19.html#sec19.5.1.

putObject_expectedBucketOwner :: Lens' PutObject (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putObject_serverSideEncryption :: Lens' PutObject (Maybe ServerSideEncryption)

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

putObject_contentType :: Lens' PutObject (Maybe Text)

A standard MIME type describing the format of the contents. For more information, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.17.

putObject_bucket :: Lens' PutObject BucketName

The bucket name to which the PUT action was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

putObject_key :: Lens' PutObject ObjectKey

Object key for which the PUT action was initiated.

putObject_body :: Lens' PutObject RequestBody

Object data.

Destructuring the Response

data PutObjectResponse

See: newPutObjectResponse smart constructor.

Constructors

PutObjectResponse'

Fields

requestCharged :: Maybe RequestCharged

eTag :: Maybe ETag

Entity tag for the uploaded object.

versionId :: Maybe ObjectVersionId

Version of the object.

expiration :: Maybe Text

If the expiration is configured for the object (see PutBucketLifecycleConfiguration), the response includes this header. It includes the expiry-date and rule-id key-value pairs that provide information about object expiration. The value of the rule-id is URL encoded.

sSECustomerAlgorithm :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

bucketKeyEnabled :: Maybe Bool

Indicates whether the uploaded object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

sSECustomerKeyMD5 :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

sSEKMSKeyId :: Maybe (Sensitive Text)

If x-amz-server-side-encryption is present and has the value of aws:kms, this header specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

sSEKMSEncryptionContext :: Maybe (Sensitive Text)

If present, specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

serverSideEncryption :: Maybe ServerSideEncryption

If you specified server-side encryption either with an Amazon Web Services KMS key or Amazon S3-managed encryption key in your PUT request, the response includes this header. It confirms the encryption algorithm that Amazon S3 used to encrypt the object.

httpStatus :: Int

The response's http status code.

newPutObjectResponse

Arguments

:: Int
$sel:httpStatus:PutObjectResponse'

-> PutObjectResponse

Create a value of PutObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharged:PutObjectResponse', putObjectResponse_requestCharged - Undocumented member.

$sel:eTag:PutObjectResponse', putObjectResponse_eTag - Entity tag for the uploaded object.

$sel:versionId:PutObjectResponse', putObjectResponse_versionId - Version of the object.

$sel:expiration:PutObjectResponse', putObjectResponse_expiration - If the expiration is configured for the object (see PutBucketLifecycleConfiguration), the response includes this header. It includes the expiry-date and rule-id key-value pairs that provide information about object expiration. The value of the rule-id is URL encoded.

$sel:sSECustomerAlgorithm:PutObject', putObjectResponse_sSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

$sel:bucketKeyEnabled:PutObject', putObjectResponse_bucketKeyEnabled - Indicates whether the uploaded object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

$sel:sSECustomerKeyMD5:PutObject', putObjectResponse_sSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

$sel:sSEKMSKeyId:PutObject', putObjectResponse_sSEKMSKeyId - If x-amz-server-side-encryption is present and has the value of aws:kms, this header specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

$sel:sSEKMSEncryptionContext:PutObject', putObjectResponse_sSEKMSEncryptionContext - If present, specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

$sel:serverSideEncryption:PutObject', putObjectResponse_serverSideEncryption - If you specified server-side encryption either with an Amazon Web Services KMS key or Amazon S3-managed encryption key in your PUT request, the response includes this header. It confirms the encryption algorithm that Amazon S3 used to encrypt the object.

$sel:httpStatus:PutObjectResponse', putObjectResponse_httpStatus - The response's http status code.

Response Lenses

putObjectResponse_requestCharged :: Lens' PutObjectResponse (Maybe RequestCharged)

Undocumented member.

putObjectResponse_eTag :: Lens' PutObjectResponse (Maybe ETag)

Entity tag for the uploaded object.

putObjectResponse_versionId :: Lens' PutObjectResponse (Maybe ObjectVersionId)

Version of the object.

putObjectResponse_expiration :: Lens' PutObjectResponse (Maybe Text)

If the expiration is configured for the object (see PutBucketLifecycleConfiguration), the response includes this header. It includes the expiry-date and rule-id key-value pairs that provide information about object expiration. The value of the rule-id is URL encoded.

putObjectResponse_sSECustomerAlgorithm :: Lens' PutObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

putObjectResponse_bucketKeyEnabled :: Lens' PutObjectResponse (Maybe Bool)

Indicates whether the uploaded object uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

putObjectResponse_sSECustomerKeyMD5 :: Lens' PutObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

putObjectResponse_sSEKMSKeyId :: Lens' PutObjectResponse (Maybe Text)

If x-amz-server-side-encryption is present and has the value of aws:kms, this header specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

putObjectResponse_sSEKMSEncryptionContext :: Lens' PutObjectResponse (Maybe Text)

If present, specifies the Amazon Web Services KMS Encryption Context to use for object encryption. The value of this header is a base64-encoded UTF-8 string holding JSON with the encryption context key-value pairs.

putObjectResponse_serverSideEncryption :: Lens' PutObjectResponse (Maybe ServerSideEncryption)

If you specified server-side encryption either with an Amazon Web Services KMS key or Amazon S3-managed encryption key in your PUT request, the response includes this header. It confirms the encryption algorithm that Amazon S3 used to encrypt the object.

putObjectResponse_httpStatus :: Lens' PutObjectResponse Int

The response's http status code.