Amazonka.S3.PutObjectLockConfiguration

Description

Places an Object Lock configuration on the specified bucket. The rule specified in the Object Lock configuration will be applied by default to every new object placed in the specified bucket. For more information, see Locking Objects.

The DefaultRetention settings require both a mode and a period.

The DefaultRetention period can be either Days or Years but you must select one. You cannot specify Days and Years at the same time.

You can only enable Object Lock for new buckets. If you want to turn on Object Lock for an existing bucket, contact Amazon Web Services Support.

Creating a Request

data PutObjectLockConfiguration

See: newPutObjectLockConfiguration smart constructor.

Constructors

PutObjectLockConfiguration'

Fields

token :: Maybe Text

A token to allow Object Lock to be enabled for an existing bucket.

objectLockConfiguration :: Maybe ObjectLockConfiguration

The Object Lock configuration that you want to apply to the specified bucket.

requestPayer :: Maybe RequestPayer

contentMD5 :: Maybe Text

The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket whose Object Lock configuration you want to create or replace.

newPutObjectLockConfiguration

Arguments

:: BucketName
$sel:bucket:PutObjectLockConfiguration'

-> PutObjectLockConfiguration

Create a value of PutObjectLockConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:token:PutObjectLockConfiguration', putObjectLockConfiguration_token - A token to allow Object Lock to be enabled for an existing bucket.

$sel:objectLockConfiguration:PutObjectLockConfiguration', putObjectLockConfiguration_objectLockConfiguration - The Object Lock configuration that you want to apply to the specified bucket.

$sel:requestPayer:PutObjectLockConfiguration', putObjectLockConfiguration_requestPayer - Undocumented member.

$sel:contentMD5:PutObjectLockConfiguration', putObjectLockConfiguration_contentMD5 - The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

$sel:expectedBucketOwner:PutObjectLockConfiguration', putObjectLockConfiguration_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:PutObjectLockConfiguration', putObjectLockConfiguration_bucket - The bucket whose Object Lock configuration you want to create or replace.

Request Lenses

putObjectLockConfiguration_token :: Lens' PutObjectLockConfiguration (Maybe Text)

A token to allow Object Lock to be enabled for an existing bucket.

putObjectLockConfiguration_objectLockConfiguration :: Lens' PutObjectLockConfiguration (Maybe ObjectLockConfiguration)

The Object Lock configuration that you want to apply to the specified bucket.

putObjectLockConfiguration_requestPayer :: Lens' PutObjectLockConfiguration (Maybe RequestPayer)

Undocumented member.

putObjectLockConfiguration_contentMD5 :: Lens' PutObjectLockConfiguration (Maybe Text)

The MD5 hash for the request body.

For requests made using the Amazon Web Services Command Line Interface (CLI) or Amazon Web Services SDKs, this field is calculated automatically.

putObjectLockConfiguration_expectedBucketOwner :: Lens' PutObjectLockConfiguration (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

putObjectLockConfiguration_bucket :: Lens' PutObjectLockConfiguration BucketName

The bucket whose Object Lock configuration you want to create or replace.

Destructuring the Response

data PutObjectLockConfigurationResponse

See: newPutObjectLockConfigurationResponse smart constructor.

Constructors

PutObjectLockConfigurationResponse'

Fields

requestCharged :: Maybe RequestCharged

httpStatus :: Int

The response's http status code.

newPutObjectLockConfigurationResponse

Arguments

:: Int
$sel:httpStatus:PutObjectLockConfigurationResponse'

-> PutObjectLockConfigurationResponse

Create a value of PutObjectLockConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharged:PutObjectLockConfigurationResponse', putObjectLockConfigurationResponse_requestCharged - Undocumented member.

$sel:httpStatus:PutObjectLockConfigurationResponse', putObjectLockConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

putObjectLockConfigurationResponse_requestCharged :: Lens' PutObjectLockConfigurationResponse (Maybe RequestCharged)

Undocumented member.

putObjectLockConfigurationResponse_httpStatus :: Lens' PutObjectLockConfigurationResponse Int

The response's http status code.