Amazonka.S3.Types.CSVOutput

Description

Documentation

data CSVOutput

Describes how uncompressed comma-separated values (CSV)-formatted results are formatted.

See: newCSVOutput smart constructor.

Constructors

CSVOutput'

Fields

quoteCharacter :: Maybe Text

A single character used for escaping when the field delimiter is part of the value. For example, if the value is a, b, Amazon S3 wraps this field value in quotation marks, as follows: " a , b ".

quoteFields :: Maybe QuoteFields

Indicates whether to use quotation marks around output fields.

ALWAYS: Always use quotation marks for output fields.

ASNEEDED: Use quotation marks for output fields when needed.

recordDelimiter :: Maybe Text

A single character used to separate individual records in the output. Instead of the default value, you can specify an arbitrary delimiter.

quoteEscapeCharacter :: Maybe Text

The single character used for escaping the quote character inside an already escaped value.

fieldDelimiter :: Maybe Text

The value used to separate individual fields in a record. You can specify an arbitrary delimiter.

newCSVOutput :: CSVOutput

Create a value of CSVOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quoteCharacter:CSVOutput', cSVOutput_quoteCharacter - A single character used for escaping when the field delimiter is part of the value. For example, if the value is a, b, Amazon S3 wraps this field value in quotation marks, as follows: " a , b ".

$sel:quoteFields:CSVOutput', cSVOutput_quoteFields - Indicates whether to use quotation marks around output fields.

ALWAYS: Always use quotation marks for output fields.

ASNEEDED: Use quotation marks for output fields when needed.

$sel:recordDelimiter:CSVOutput', cSVOutput_recordDelimiter - A single character used to separate individual records in the output. Instead of the default value, you can specify an arbitrary delimiter.

$sel:quoteEscapeCharacter:CSVOutput', cSVOutput_quoteEscapeCharacter - The single character used for escaping the quote character inside an already escaped value.

$sel:fieldDelimiter:CSVOutput', cSVOutput_fieldDelimiter - The value used to separate individual fields in a record. You can specify an arbitrary delimiter.

cSVOutput_quoteCharacter :: Lens' CSVOutput (Maybe Text)

A single character used for escaping when the field delimiter is part of the value. For example, if the value is a, b, Amazon S3 wraps this field value in quotation marks, as follows: " a , b ".

cSVOutput_quoteFields :: Lens' CSVOutput (Maybe QuoteFields)

Indicates whether to use quotation marks around output fields.

ALWAYS: Always use quotation marks for output fields.

ASNEEDED: Use quotation marks for output fields when needed.

cSVOutput_recordDelimiter :: Lens' CSVOutput (Maybe Text)

A single character used to separate individual records in the output. Instead of the default value, you can specify an arbitrary delimiter.

cSVOutput_quoteEscapeCharacter :: Lens' CSVOutput (Maybe Text)

The single character used for escaping the quote character inside an already escaped value.

cSVOutput_fieldDelimiter :: Lens' CSVOutput (Maybe Text)

The value used to separate individual fields in a record. You can specify an arbitrary delimiter.