Amazonka.S3.Types.EndEvent

Description

Documentation

data EndEvent

A message that indicates the request is complete and no more messages will be sent. You should not assume that the request is complete until the client receives an EndEvent.

See: newEndEvent smart constructor.

Constructors

EndEvent'

newEndEvent :: EndEvent

Create a value of EndEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.