Amazonka.S3.Types.FilterRule

Description

Documentation

data FilterRule

Specifies the Amazon S3 object key name to filter on and whether to filter on the suffix or prefix of the key name.

See: newFilterRule smart constructor.

Constructors

FilterRule'

Fields

value :: Maybe Text

The value that the filter searches for in object key names.

name :: Maybe FilterRuleName

The object key name prefix or suffix identifying one or more objects to which the filtering rule applies. The maximum length is 1,024 characters. Overlapping prefixes and suffixes are not supported. For more information, see Configuring Event Notifications in the Amazon S3 User Guide.

newFilterRule :: FilterRule

Create a value of FilterRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:FilterRule', filterRule_value - The value that the filter searches for in object key names.

$sel:name:FilterRule', filterRule_name - The object key name prefix or suffix identifying one or more objects to which the filtering rule applies. The maximum length is 1,024 characters. Overlapping prefixes and suffixes are not supported. For more information, see Configuring Event Notifications in the Amazon S3 User Guide.

filterRule_value :: Lens' FilterRule (Maybe Text)

The value that the filter searches for in object key names.

filterRule_name :: Lens' FilterRule (Maybe FilterRuleName)

The object key name prefix or suffix identifying one or more objects to which the filtering rule applies. The maximum length is 1,024 characters. Overlapping prefixes and suffixes are not supported. For more information, see Configuring Event Notifications in the Amazon S3 User Guide.