Amazonka.S3.Types.IntelligentTieringConfiguration

Description

Documentation

data IntelligentTieringConfiguration

Specifies the S3 Intelligent-Tiering configuration for an Amazon S3 bucket.

For information about the S3 Intelligent-Tiering storage class, see Storage class for automatically optimizing frequently and infrequently accessed objects.

See: newIntelligentTieringConfiguration smart constructor.

Constructors

IntelligentTieringConfiguration'

Fields

filter' :: Maybe IntelligentTieringFilter

Specifies a bucket filter. The configuration only includes objects that meet the filter's criteria.

id :: Text

The ID used to identify the S3 Intelligent-Tiering configuration.

status :: IntelligentTieringStatus

Specifies the status of the configuration.

tierings :: [Tiering]

Specifies the S3 Intelligent-Tiering storage class tier of the configuration.

newIntelligentTieringConfiguration

Arguments

:: Text
$sel:id:IntelligentTieringConfiguration'

-> IntelligentTieringStatus
$sel:status:IntelligentTieringConfiguration'

-> IntelligentTieringConfiguration

Create a value of IntelligentTieringConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filter':IntelligentTieringConfiguration', intelligentTieringConfiguration_filter - Specifies a bucket filter. The configuration only includes objects that meet the filter's criteria.

$sel:id:IntelligentTieringConfiguration', intelligentTieringConfiguration_id - The ID used to identify the S3 Intelligent-Tiering configuration.

$sel:status:IntelligentTieringConfiguration', intelligentTieringConfiguration_status - Specifies the status of the configuration.

$sel:tierings:IntelligentTieringConfiguration', intelligentTieringConfiguration_tierings - Specifies the S3 Intelligent-Tiering storage class tier of the configuration.

intelligentTieringConfiguration_filter :: Lens' IntelligentTieringConfiguration (Maybe IntelligentTieringFilter)

Specifies a bucket filter. The configuration only includes objects that meet the filter's criteria.

intelligentTieringConfiguration_id :: Lens' IntelligentTieringConfiguration Text

The ID used to identify the S3 Intelligent-Tiering configuration.

intelligentTieringConfiguration_status :: Lens' IntelligentTieringConfiguration IntelligentTieringStatus

Specifies the status of the configuration.

intelligentTieringConfiguration_tierings :: Lens' IntelligentTieringConfiguration [Tiering]

Specifies the S3 Intelligent-Tiering storage class tier of the configuration.