Amazonka.S3.Types.MetricsFilter

Description

Documentation

data MetricsFilter

Specifies a metrics configuration filter. The metrics configuration only includes objects that meet the filter's criteria. A filter must be a prefix, an object tag, an access point ARN, or a conjunction (MetricsAndOperator). For more information, see PutBucketMetricsConfiguration.

See: newMetricsFilter smart constructor.

Constructors

MetricsFilter'

Fields

tag :: Maybe Tag

The tag used when evaluating a metrics filter.

prefix :: Maybe Text

The prefix used when evaluating a metrics filter.

and :: Maybe MetricsAndOperator

A conjunction (logical AND) of predicates, which is used in evaluating a metrics filter. The operator must have at least two predicates, and an object must match all of the predicates in order for the filter to apply.

accessPointArn :: Maybe Text

The access point ARN used when evaluating a metrics filter.

newMetricsFilter :: MetricsFilter

Create a value of MetricsFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tag:MetricsFilter', metricsFilter_tag - The tag used when evaluating a metrics filter.

$sel:prefix:MetricsFilter', metricsFilter_prefix - The prefix used when evaluating a metrics filter.

$sel:and:MetricsFilter', metricsFilter_and - A conjunction (logical AND) of predicates, which is used in evaluating a metrics filter. The operator must have at least two predicates, and an object must match all of the predicates in order for the filter to apply.

$sel:accessPointArn:MetricsFilter', metricsFilter_accessPointArn - The access point ARN used when evaluating a metrics filter.

metricsFilter_tag :: Lens' MetricsFilter (Maybe Tag)

The tag used when evaluating a metrics filter.

metricsFilter_prefix :: Lens' MetricsFilter (Maybe Text)

The prefix used when evaluating a metrics filter.

metricsFilter_and :: Lens' MetricsFilter (Maybe MetricsAndOperator)

A conjunction (logical AND) of predicates, which is used in evaluating a metrics filter. The operator must have at least two predicates, and an object must match all of the predicates in order for the filter to apply.

metricsFilter_accessPointArn :: Lens' MetricsFilter (Maybe Text)

The access point ARN used when evaluating a metrics filter.