Amazonka.S3.Types.Object

Description

Documentation

data Object

An object consists of data and its descriptive metadata.

See: newObject smart constructor.

Constructors

Object'

Fields

owner :: Maybe Owner

The owner of the object

eTag :: ETag

The entity tag is a hash of the object. The ETag reflects changes only to the contents of an object, not its metadata. The ETag may or may not be an MD5 digest of the object data. Whether or not it is depends on how the object was created and how it is encrypted as described below:

Objects created by the PUT Object, POST Object, or Copy operation, or through the Amazon Web Services Management Console, and are encrypted by SSE-S3 or plaintext, have ETags that are an MD5 digest of their object data.

Objects created by the PUT Object, POST Object, or Copy operation, or through the Amazon Web Services Management Console, and are encrypted by SSE-C or SSE-KMS, have ETags that are not an MD5 digest of their object data.

If an object is created by either the Multipart Upload or Part Copy operation, the ETag is not an MD5 digest, regardless of the method of encryption.

size :: Integer

Size in bytes of the object

key :: ObjectKey

The name that you assign to an object. You use the object key to retrieve the object.

storageClass :: ObjectStorageClass

The class of storage used to store the object.

lastModified :: ISO8601

Creation date of the object.

newObject

Arguments

:: ETag
$sel:eTag:Object'

-> Integer
$sel:size:Object'

-> ObjectKey
$sel:key:Object'

-> ObjectStorageClass
$sel:storageClass:Object'

-> UTCTime
$sel:lastModified:Object'

-> Object

Create a value of Object with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:owner:Object', object_owner - The owner of the object

$sel:eTag:Object', object_eTag - The entity tag is a hash of the object. The ETag reflects changes only to the contents of an object, not its metadata. The ETag may or may not be an MD5 digest of the object data. Whether or not it is depends on how the object was created and how it is encrypted as described below:

Objects created by the PUT Object, POST Object, or Copy operation, or through the Amazon Web Services Management Console, and are encrypted by SSE-S3 or plaintext, have ETags that are an MD5 digest of their object data.

Objects created by the PUT Object, POST Object, or Copy operation, or through the Amazon Web Services Management Console, and are encrypted by SSE-C or SSE-KMS, have ETags that are not an MD5 digest of their object data.

If an object is created by either the Multipart Upload or Part Copy operation, the ETag is not an MD5 digest, regardless of the method of encryption.

$sel:size:Object', object_size - Size in bytes of the object

$sel:key:Object', object_key - The name that you assign to an object. You use the object key to retrieve the object.

$sel:storageClass:Object', object_storageClass - The class of storage used to store the object.

$sel:lastModified:Object', object_lastModified - Creation date of the object.

object_owner :: Lens' Object (Maybe Owner)

The owner of the object

object_eTag :: Lens' Object ETag

The entity tag is a hash of the object. The ETag reflects changes only to the contents of an object, not its metadata. The ETag may or may not be an MD5 digest of the object data. Whether or not it is depends on how the object was created and how it is encrypted as described below:

Objects created by the PUT Object, POST Object, or Copy operation, or through the Amazon Web Services Management Console, and are encrypted by SSE-S3 or plaintext, have ETags that are an MD5 digest of their object data.

Objects created by the PUT Object, POST Object, or Copy operation, or through the Amazon Web Services Management Console, and are encrypted by SSE-C or SSE-KMS, have ETags that are not an MD5 digest of their object data.

If an object is created by either the Multipart Upload or Part Copy operation, the ETag is not an MD5 digest, regardless of the method of encryption.

object_size :: Lens' Object Integer

Size in bytes of the object

object_key :: Lens' Object ObjectKey

The name that you assign to an object. You use the object key to retrieve the object.

object_storageClass :: Lens' Object ObjectStorageClass

The class of storage used to store the object.

object_lastModified :: Lens' Object UTCTime

Creation date of the object.