Amazonka.S3.Types.ObjectLockConfiguration

Description

Documentation

data ObjectLockConfiguration

The container element for Object Lock configuration parameters.

See: newObjectLockConfiguration smart constructor.

Constructors

ObjectLockConfiguration'

Fields

objectLockEnabled :: Maybe ObjectLockEnabled

Indicates whether this bucket has an Object Lock configuration enabled. Enable ObjectLockEnabled when you apply ObjectLockConfiguration to a bucket.

rule :: Maybe ObjectLockRule

Specifies the Object Lock rule for the specified object. Enable the this rule when you apply ObjectLockConfiguration to a bucket. Bucket settings require both a mode and a period. The period can be either Days or Years but you must select one. You cannot specify Days and Years at the same time.

newObjectLockConfiguration :: ObjectLockConfiguration

Create a value of ObjectLockConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:objectLockEnabled:ObjectLockConfiguration', objectLockConfiguration_objectLockEnabled - Indicates whether this bucket has an Object Lock configuration enabled. Enable ObjectLockEnabled when you apply ObjectLockConfiguration to a bucket.

$sel:rule:ObjectLockConfiguration', objectLockConfiguration_rule - Specifies the Object Lock rule for the specified object. Enable the this rule when you apply ObjectLockConfiguration to a bucket. Bucket settings require both a mode and a period. The period can be either Days or Years but you must select one. You cannot specify Days and Years at the same time.

objectLockConfiguration_objectLockEnabled :: Lens' ObjectLockConfiguration (Maybe ObjectLockEnabled)

Indicates whether this bucket has an Object Lock configuration enabled. Enable ObjectLockEnabled when you apply ObjectLockConfiguration to a bucket.

objectLockConfiguration_rule :: Lens' ObjectLockConfiguration (Maybe ObjectLockRule)

Specifies the Object Lock rule for the specified object. Enable the this rule when you apply ObjectLockConfiguration to a bucket. Bucket settings require both a mode and a period. The period can be either Days or Years but you must select one. You cannot specify Days and Years at the same time.