Amazonka.S3.Types.ObjectLockRule

Description

Documentation

data ObjectLockRule

The container element for an Object Lock rule.

See: newObjectLockRule smart constructor.

Constructors

ObjectLockRule'

Fields

defaultRetention :: Maybe DefaultRetention

The default Object Lock retention mode and period that you want to apply to new objects placed in the specified bucket. Bucket settings require both a mode and a period. The period can be either Days or Years but you must select one. You cannot specify Days and Years at the same time.

newObjectLockRule :: ObjectLockRule

Create a value of ObjectLockRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultRetention:ObjectLockRule', objectLockRule_defaultRetention - The default Object Lock retention mode and period that you want to apply to new objects placed in the specified bucket. Bucket settings require both a mode and a period. The period can be either Days or Years but you must select one. You cannot specify Days and Years at the same time.

objectLockRule_defaultRetention :: Lens' ObjectLockRule (Maybe DefaultRetention)

The default Object Lock retention mode and period that you want to apply to new objects placed in the specified bucket. Bucket settings require both a mode and a period. The period can be either Days or Years but you must select one. You cannot specify Days and Years at the same time.