Amazonka.S3.Types.OutputLocation

Description

Documentation

data OutputLocation

Describes the location where the restore job's output is stored.

See: newOutputLocation smart constructor.

Constructors

OutputLocation'

Fields

s3 :: Maybe S3Location

Describes an S3 location that will receive the results of the restore request.

newOutputLocation :: OutputLocation

Create a value of OutputLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3:OutputLocation', outputLocation_s3 - Describes an S3 location that will receive the results of the restore request.

outputLocation_s3 :: Lens' OutputLocation (Maybe S3Location)

Describes an S3 location that will receive the results of the restore request.