Amazonka.S3.Types.RestoreRequest

Description

Documentation

data RestoreRequest

Container for restore job parameters.

See: newRestoreRequest smart constructor.

Constructors

RestoreRequest'

Fields

days :: Maybe Int

Lifetime of the active copy in days. Do not use with restores that specify OutputLocation.

The Days element is required for regular restores, and must not be provided for select requests.

selectParameters :: Maybe SelectParameters

Describes the parameters for Select job types.

outputLocation :: Maybe OutputLocation

Describes the location where the restore job's output is stored.

tier :: Maybe Tier

Retrieval tier at which the restore will be processed.

glacierJobParameters :: Maybe GlacierJobParameters

S3 Glacier related parameters pertaining to this job. Do not use with restores that specify OutputLocation.

type' :: Maybe RestoreRequestType

Type of restore request.

description :: Maybe Text

The optional description for the job.

newRestoreRequest :: RestoreRequest

Create a value of RestoreRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:days:RestoreRequest', restoreRequest_days - Lifetime of the active copy in days. Do not use with restores that specify OutputLocation.

The Days element is required for regular restores, and must not be provided for select requests.

$sel:selectParameters:RestoreRequest', restoreRequest_selectParameters - Describes the parameters for Select job types.

$sel:outputLocation:RestoreRequest', restoreRequest_outputLocation - Describes the location where the restore job's output is stored.

$sel:tier:RestoreRequest', restoreRequest_tier - Retrieval tier at which the restore will be processed.

$sel:glacierJobParameters:RestoreRequest', restoreRequest_glacierJobParameters - S3 Glacier related parameters pertaining to this job. Do not use with restores that specify OutputLocation.

$sel:type':RestoreRequest', restoreRequest_type - Type of restore request.

$sel:description:RestoreRequest', restoreRequest_description - The optional description for the job.

restoreRequest_days :: Lens' RestoreRequest (Maybe Int)

Lifetime of the active copy in days. Do not use with restores that specify OutputLocation.

The Days element is required for regular restores, and must not be provided for select requests.

restoreRequest_selectParameters :: Lens' RestoreRequest (Maybe SelectParameters)

Describes the parameters for Select job types.

restoreRequest_outputLocation :: Lens' RestoreRequest (Maybe OutputLocation)

Describes the location where the restore job's output is stored.

restoreRequest_tier :: Lens' RestoreRequest (Maybe Tier)

Retrieval tier at which the restore will be processed.

restoreRequest_glacierJobParameters :: Lens' RestoreRequest (Maybe GlacierJobParameters)

S3 Glacier related parameters pertaining to this job. Do not use with restores that specify OutputLocation.

restoreRequest_type :: Lens' RestoreRequest (Maybe RestoreRequestType)

Type of restore request.

restoreRequest_description :: Lens' RestoreRequest (Maybe Text)

The optional description for the job.