Amazonka.S3.Types.WebsiteConfiguration

Description

Documentation

data WebsiteConfiguration

Specifies website configuration parameters for an Amazon S3 bucket.

See: newWebsiteConfiguration smart constructor.

Constructors

WebsiteConfiguration'

Fields

redirectAllRequestsTo :: Maybe RedirectAllRequestsTo

The redirect behavior for every request to this bucket's website endpoint.

If you specify this property, you can't specify any other property.

errorDocument :: Maybe ErrorDocument

The name of the error document for the website.

indexDocument :: Maybe IndexDocument

The name of the index document for the website.

routingRules :: Maybe [RoutingRule]

Rules that define when a redirect is applied and the redirect behavior.

newWebsiteConfiguration :: WebsiteConfiguration

Create a value of WebsiteConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:redirectAllRequestsTo:WebsiteConfiguration', websiteConfiguration_redirectAllRequestsTo - The redirect behavior for every request to this bucket's website endpoint.

If you specify this property, you can't specify any other property.

$sel:errorDocument:WebsiteConfiguration', websiteConfiguration_errorDocument - The name of the error document for the website.

$sel:indexDocument:WebsiteConfiguration', websiteConfiguration_indexDocument - The name of the index document for the website.

$sel:routingRules:WebsiteConfiguration', websiteConfiguration_routingRules - Rules that define when a redirect is applied and the redirect behavior.

websiteConfiguration_redirectAllRequestsTo :: Lens' WebsiteConfiguration (Maybe RedirectAllRequestsTo)

The redirect behavior for every request to this bucket's website endpoint.

If you specify this property, you can't specify any other property.

websiteConfiguration_errorDocument :: Lens' WebsiteConfiguration (Maybe ErrorDocument)

The name of the error document for the website.

websiteConfiguration_indexDocument :: Lens' WebsiteConfiguration (Maybe IndexDocument)

The name of the index document for the website.

websiteConfiguration_routingRules :: Lens' WebsiteConfiguration (Maybe [RoutingRule])

Rules that define when a redirect is applied and the redirect behavior.