Amazonka.S3.UploadPart

Description

Uploads a part in a multipart upload.

In this operation, you provide part data in your request. However, you have an option to specify your existing Amazon S3 object as a data source for the part you are uploading. To upload a part from an existing object, you use the UploadPartCopy operation.

You must initiate a multipart upload (see CreateMultipartUpload) before you can upload any part. In response to your initiate request, Amazon S3 returns an upload ID, a unique identifier, that you must include in your upload part request.

Part numbers can be any number from 1 to 10,000, inclusive. A part number uniquely identifies a part and also defines its position within the object being created. If you upload a new part using the same part number that was used with a previous part, the previously uploaded part is overwritten. Each part must be at least 5 MB in size, except the last part. There is no size limit on the last part of your multipart upload.

To ensure that data is not corrupted when traversing the network, specify the Content-MD5 header in the upload part request. Amazon S3 checks the part data against the provided MD5 value. If they do not match, Amazon S3 returns an error.

If the upload request is signed with Signature Version 4, then Amazon Web Services S3 uses the x-amz-content-sha256 header as a checksum instead of Content-MD5. For more information see Authenticating Requests: Using the Authorization Header (Amazon Web Services Signature Version 4).

Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

For more information on multipart uploads, go to Multipart Upload Overview in the Amazon S3 User Guide .

For information on the permissions required to use the multipart upload API, go to Multipart Upload and Permissions in the Amazon S3 User Guide.

You can optionally request server-side encryption where Amazon S3 encrypts your data as it writes it to disks in its data centers and decrypts it for you when you access it. You have the option of providing your own encryption key, or you can use the Amazon Web Services managed encryption keys. If you choose to provide your own encryption key, the request headers you provide in the request must match the headers you used in the request to initiate the upload by using CreateMultipartUpload. For more information, go to Using Server-Side Encryption in the Amazon S3 User Guide.

Server-side encryption is supported by the S3 Multipart Upload actions. Unless you are using a customer-provided encryption key, you don't need to specify the encryption parameters in each UploadPart request. Instead, you only need to specify the server-side encryption parameters in the initial Initiate Multipart request. For more information, see CreateMultipartUpload.

If you requested server-side encryption using a customer-provided encryption key in your initiate multipart upload request, you must provide identical encryption information in each part upload using the following headers.

x-amz-server-side-encryption-customer-algorithm

x-amz-server-side-encryption-customer-key

x-amz-server-side-encryption-customer-key-MD5

Special Errors

- Code: NoSuchUpload

/Cause: The specified multipart upload does not exist. The upload ID might be invalid, or the multipart upload might have been aborted or completed./

HTTP Status Code: 404 Not Found

SOAP Fault Code Prefix: Client

Related Resources

CreateMultipartUpload

CompleteMultipartUpload

AbortMultipartUpload

ListParts

ListMultipartUploads

Creating a Request

data UploadPart

See: newUploadPart smart constructor.

Constructors

UploadPart'

Fields

contentLength :: Maybe Integer

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

sSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use to when encrypting the object (for example, AES256).

sSECustomerKey :: Maybe (Sensitive Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

requestPayer :: Maybe RequestPayer

sSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

contentMD5 :: Maybe Text

The base64-encoded 128-bit MD5 digest of the part data. This parameter is auto-populated when using the command from the CLI. This parameter is required if object lock parameters are specified.

expectedBucketOwner :: Maybe Text

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The name of the bucket to which the multipart upload was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

key :: ObjectKey

Object key for which the multipart upload was initiated.

partNumber :: Int

Part number of part being uploaded. This is a positive integer between 1 and 10,000.

uploadId :: Text

Upload ID identifying the multipart upload whose part is being uploaded.

body :: RequestBody

Object data.

newUploadPart

Arguments

:: BucketName
$sel:bucket:UploadPart'

-> ObjectKey
$sel:key:UploadPart'

-> Int
$sel:partNumber:UploadPart'

-> Text
$sel:uploadId:UploadPart'

-> RequestBody
$sel:body:UploadPart'

-> UploadPart

Create a value of UploadPart with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentLength:UploadPart', uploadPart_contentLength - Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

$sel:sSECustomerAlgorithm:UploadPart', uploadPart_sSECustomerAlgorithm - Specifies the algorithm to use to when encrypting the object (for example, AES256).

$sel:sSECustomerKey:UploadPart', uploadPart_sSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

$sel:requestPayer:UploadPart', uploadPart_requestPayer - Undocumented member.

$sel:sSECustomerKeyMD5:UploadPart', uploadPart_sSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

$sel:contentMD5:UploadPart', uploadPart_contentMD5 - The base64-encoded 128-bit MD5 digest of the part data. This parameter is auto-populated when using the command from the CLI. This parameter is required if object lock parameters are specified.

$sel:expectedBucketOwner:UploadPart', uploadPart_expectedBucketOwner - The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:UploadPart', uploadPart_bucket - The name of the bucket to which the multipart upload was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:key:UploadPart', uploadPart_key - Object key for which the multipart upload was initiated.

$sel:partNumber:UploadPart', uploadPart_partNumber - Part number of part being uploaded. This is a positive integer between 1 and 10,000.

$sel:uploadId:UploadPart', uploadPart_uploadId - Upload ID identifying the multipart upload whose part is being uploaded.

$sel:body:UploadPart', uploadPart_body - Object data.

Request Lenses

uploadPart_contentLength :: Lens' UploadPart (Maybe Integer)

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

uploadPart_sSECustomerAlgorithm :: Lens' UploadPart (Maybe Text)

Specifies the algorithm to use to when encrypting the object (for example, AES256).

uploadPart_sSECustomerKey :: Lens' UploadPart (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

uploadPart_requestPayer :: Lens' UploadPart (Maybe RequestPayer)

Undocumented member.

uploadPart_sSECustomerKeyMD5 :: Lens' UploadPart (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

uploadPart_contentMD5 :: Lens' UploadPart (Maybe Text)

The base64-encoded 128-bit MD5 digest of the part data. This parameter is auto-populated when using the command from the CLI. This parameter is required if object lock parameters are specified.

uploadPart_expectedBucketOwner :: Lens' UploadPart (Maybe Text)

The account ID of the expected bucket owner. If the bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

uploadPart_bucket :: Lens' UploadPart BucketName

The name of the bucket to which the multipart upload was initiated.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

uploadPart_key :: Lens' UploadPart ObjectKey

Object key for which the multipart upload was initiated.

uploadPart_partNumber :: Lens' UploadPart Int

Part number of part being uploaded. This is a positive integer between 1 and 10,000.

uploadPart_uploadId :: Lens' UploadPart Text

Upload ID identifying the multipart upload whose part is being uploaded.

uploadPart_body :: Lens' UploadPart RequestBody

Object data.

Destructuring the Response

data UploadPartResponse

See: newUploadPartResponse smart constructor.

Constructors

UploadPartResponse'

Fields

requestCharged :: Maybe RequestCharged

eTag :: Maybe ETag

Entity tag for the uploaded object.

sSECustomerAlgorithm :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

bucketKeyEnabled :: Maybe Bool

Indicates whether the multipart upload uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

sSECustomerKeyMD5 :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

sSEKMSKeyId :: Maybe (Sensitive Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key was used for the object.

serverSideEncryption :: Maybe ServerSideEncryption

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

httpStatus :: Int

The response's http status code.

newUploadPartResponse

Arguments

:: Int
$sel:httpStatus:UploadPartResponse'

-> UploadPartResponse

Create a value of UploadPartResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharged:UploadPartResponse', uploadPartResponse_requestCharged - Undocumented member.

$sel:eTag:UploadPartResponse', uploadPartResponse_eTag - Entity tag for the uploaded object.

$sel:sSECustomerAlgorithm:UploadPart', uploadPartResponse_sSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

$sel:bucketKeyEnabled:UploadPartResponse', uploadPartResponse_bucketKeyEnabled - Indicates whether the multipart upload uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

$sel:sSECustomerKeyMD5:UploadPart', uploadPartResponse_sSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

$sel:sSEKMSKeyId:UploadPartResponse', uploadPartResponse_sSEKMSKeyId - If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key was used for the object.

$sel:serverSideEncryption:UploadPartResponse', uploadPartResponse_serverSideEncryption - The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

$sel:httpStatus:UploadPartResponse', uploadPartResponse_httpStatus - The response's http status code.

Response Lenses

uploadPartResponse_requestCharged :: Lens' UploadPartResponse (Maybe RequestCharged)

Undocumented member.

uploadPartResponse_eTag :: Lens' UploadPartResponse (Maybe ETag)

Entity tag for the uploaded object.

uploadPartResponse_sSECustomerAlgorithm :: Lens' UploadPartResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

uploadPartResponse_bucketKeyEnabled :: Lens' UploadPartResponse (Maybe Bool)

Indicates whether the multipart upload uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

uploadPartResponse_sSECustomerKeyMD5 :: Lens' UploadPartResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

uploadPartResponse_sSEKMSKeyId :: Lens' UploadPartResponse (Maybe Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key was used for the object.

uploadPartResponse_serverSideEncryption :: Lens' UploadPartResponse (Maybe ServerSideEncryption)

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

uploadPartResponse_httpStatus :: Lens' UploadPartResponse Int

The response's http status code.