Amazonka.S3.UploadPartCopy

Description

Uploads a part by copying data from an existing object as data source. You specify the data source by adding the request header x-amz-copy-source in your request and a byte range by adding the request header x-amz-copy-source-range in your request.

The minimum allowable part size for a multipart upload is 5 MB. For more information about multipart upload limits, go to Quick Facts in the Amazon S3 User Guide.

Instead of using an existing object as part data, you might use the UploadPart action and provide data in your request.

You must initiate a multipart upload before you can upload any part. In response to your initiate request. Amazon S3 returns a unique identifier, the upload ID, that you must include in your upload part request.

For more information about using the UploadPartCopy operation, see the following:

For conceptual information about multipart uploads, see Uploading Objects Using Multipart Upload in the Amazon S3 User Guide.

For information about permissions required to use the multipart upload API, see Multipart Upload and Permissions in the Amazon S3 User Guide.

For information about copying objects using a single atomic action vs. the multipart upload, see Operations on Objects in the Amazon S3 User Guide.

For information about using server-side encryption with customer-provided encryption keys with the UploadPartCopy operation, see CopyObject and UploadPart.

Note the following additional considerations about the request headers x-amz-copy-source-if-match, x-amz-copy-source-if-none-match, x-amz-copy-source-if-unmodified-since, and x-amz-copy-source-if-modified-since:

Consideration 1 - If both of the x-amz-copy-source-if-match and x-amz-copy-source-if-unmodified-since headers are present in the request as follows:

x-amz-copy-source-if-match condition evaluates to true, and;

x-amz-copy-source-if-unmodified-since condition evaluates to false;

Amazon S3 returns 200 OK and copies the data.

Consideration 2 - If both of the x-amz-copy-source-if-none-match and x-amz-copy-source-if-modified-since headers are present in the request as follows:

x-amz-copy-source-if-none-match condition evaluates to false, and;

x-amz-copy-source-if-modified-since condition evaluates to true;

Amazon S3 returns 412 Precondition Failed response code.

Versioning

If your bucket has versioning enabled, you could have multiple versions of the same object. By default, x-amz-copy-source identifies the current version of the object to copy. If the current version is a delete marker and you don't specify a versionId in the x-amz-copy-source, Amazon S3 returns a 404 error, because the object does not exist. If you specify versionId in the x-amz-copy-source and the versionId is a delete marker, Amazon S3 returns an HTTP 400 error, because you are not allowed to specify a delete marker as a version for the x-amz-copy-source.

You can optionally specify a specific version of the source object to copy by adding the versionId subresource as shown in the following example:

x-amz-copy-source: /bucket/object?versionId=version id

Special Errors

- Code: NoSuchUpload

/Cause: The specified multipart upload does not exist. The upload ID might be invalid, or the multipart upload might have been aborted or completed./

HTTP Status Code: 404 Not Found

- Code: InvalidRequest

/Cause: The specified copy source is not supported as a byte-range copy source./

HTTP Status Code: 400 Bad Request

Related Resources

CreateMultipartUpload

UploadPart

CompleteMultipartUpload

AbortMultipartUpload

ListParts

ListMultipartUploads

Creating a Request

data UploadPartCopy

See: newUploadPartCopy smart constructor.

Constructors

UploadPartCopy'

Fields

copySourceIfModifiedSince :: Maybe ISO8601

Copies the object if it has been modified since the specified time.

copySourceIfUnmodifiedSince :: Maybe ISO8601

Copies the object if it hasn't been modified since the specified time.

copySourceRange :: Maybe Text

The range of bytes to copy from the source object. The range value must use the form bytes=first-last, where the first and last are the zero-based byte offsets to copy. For example, bytes=0-9 indicates that you want to copy the first 10 bytes of the source. You can copy a range only if the source object is greater than 5 MB.

copySourceSSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

copySourceIfNoneMatch :: Maybe Text

Copies the object if its entity tag (ETag) is different than the specified ETag.

sSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use to when encrypting the object (for example, AES256).

sSECustomerKey :: Maybe (Sensitive Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

requestPayer :: Maybe RequestPayer

copySourceIfMatch :: Maybe Text

Copies the object if its entity tag (ETag) matches the specified tag.

expectedSourceBucketOwner :: Maybe Text

The account ID of the expected source bucket owner. If the source bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

sSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

copySourceSSECustomerKey :: Maybe (Sensitive Text)

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

copySourceSSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use when decrypting the source object (for example, AES256).

expectedBucketOwner :: Maybe Text

The account ID of the expected destination bucket owner. If the destination bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

bucket :: BucketName

The bucket name.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

copySource :: Text

Specifies the source object for the copy operation. You specify the value in one of two formats, depending on whether you want to access the source object through an access point:

For objects not accessed through an access point, specify the name of the source bucket and key of the source object, separated by a slash (/). For example, to copy the object reports/january.pdf from the bucket awsexamplebucket, use awsexamplebucket/reports/january.pdf. The value must be URL encoded.

For objects accessed through access points, specify the Amazon Resource Name (ARN) of the object as accessed through the access point, in the format arn:aws:s3:<Region>:<account-id>:accesspoint/<access-point-name>/object/<key>. For example, to copy the object reports/january.pdf through access point my-access-point owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3:us-west-2:123456789012:accesspoint/my-access-point/object/reports/january.pdf. The value must be URL encoded.

Amazon S3 supports copy operations using access points only when the source and destination buckets are in the same Amazon Web Services Region.

Alternatively, for objects accessed through Amazon S3 on Outposts, specify the ARN of the object as accessed in the format arn:aws:s3-outposts:<Region>:<account-id>:outpost/<outpost-id>/object/<key>. For example, to copy the object reports/january.pdf through outpost my-outpost owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3-outposts:us-west-2:123456789012:outpost/my-outpost/object/reports/january.pdf. The value must be URL encoded.

To copy a specific version of an object, append ?versionId=<version-id> to the value (for example, awsexamplebucket/reports/january.pdf?versionId=QUpfdndhfd8438MNFDN93jdnJFkdmqnh893). If you don't specify a version ID, Amazon S3 copies the latest version of the source object.

key :: ObjectKey

Object key for which the multipart upload was initiated.

partNumber :: Int

Part number of part being copied. This is a positive integer between 1 and 10,000.

uploadId :: Text

Upload ID identifying the multipart upload whose part is being copied.

newUploadPartCopy

Arguments

:: BucketName
$sel:bucket:UploadPartCopy'

-> Text
$sel:copySource:UploadPartCopy'

-> ObjectKey
$sel:key:UploadPartCopy'

-> Int
$sel:partNumber:UploadPartCopy'

-> Text
$sel:uploadId:UploadPartCopy'

-> UploadPartCopy

Create a value of UploadPartCopy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:copySourceIfModifiedSince:UploadPartCopy', uploadPartCopy_copySourceIfModifiedSince - Copies the object if it has been modified since the specified time.

$sel:copySourceIfUnmodifiedSince:UploadPartCopy', uploadPartCopy_copySourceIfUnmodifiedSince - Copies the object if it hasn't been modified since the specified time.

$sel:copySourceRange:UploadPartCopy', uploadPartCopy_copySourceRange - The range of bytes to copy from the source object. The range value must use the form bytes=first-last, where the first and last are the zero-based byte offsets to copy. For example, bytes=0-9 indicates that you want to copy the first 10 bytes of the source. You can copy a range only if the source object is greater than 5 MB.

$sel:copySourceSSECustomerKeyMD5:UploadPartCopy', uploadPartCopy_copySourceSSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

$sel:copySourceIfNoneMatch:UploadPartCopy', uploadPartCopy_copySourceIfNoneMatch - Copies the object if its entity tag (ETag) is different than the specified ETag.

$sel:sSECustomerAlgorithm:UploadPartCopy', uploadPartCopy_sSECustomerAlgorithm - Specifies the algorithm to use to when encrypting the object (for example, AES256).

$sel:sSECustomerKey:UploadPartCopy', uploadPartCopy_sSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

$sel:requestPayer:UploadPartCopy', uploadPartCopy_requestPayer - Undocumented member.

$sel:copySourceIfMatch:UploadPartCopy', uploadPartCopy_copySourceIfMatch - Copies the object if its entity tag (ETag) matches the specified tag.

$sel:expectedSourceBucketOwner:UploadPartCopy', uploadPartCopy_expectedSourceBucketOwner - The account ID of the expected source bucket owner. If the source bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:sSECustomerKeyMD5:UploadPartCopy', uploadPartCopy_sSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

$sel:copySourceSSECustomerKey:UploadPartCopy', uploadPartCopy_copySourceSSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

$sel:copySourceSSECustomerAlgorithm:UploadPartCopy', uploadPartCopy_copySourceSSECustomerAlgorithm - Specifies the algorithm to use when decrypting the source object (for example, AES256).

$sel:expectedBucketOwner:UploadPartCopy', uploadPartCopy_expectedBucketOwner - The account ID of the expected destination bucket owner. If the destination bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

$sel:bucket:UploadPartCopy', uploadPartCopy_bucket - The bucket name.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

$sel:copySource:UploadPartCopy', uploadPartCopy_copySource - Specifies the source object for the copy operation. You specify the value in one of two formats, depending on whether you want to access the source object through an access point:

For objects not accessed through an access point, specify the name of the source bucket and key of the source object, separated by a slash (/). For example, to copy the object reports/january.pdf from the bucket awsexamplebucket, use awsexamplebucket/reports/january.pdf. The value must be URL encoded.

For objects accessed through access points, specify the Amazon Resource Name (ARN) of the object as accessed through the access point, in the format arn:aws:s3:<Region>:<account-id>:accesspoint/<access-point-name>/object/<key>. For example, to copy the object reports/january.pdf through access point my-access-point owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3:us-west-2:123456789012:accesspoint/my-access-point/object/reports/january.pdf. The value must be URL encoded.

Amazon S3 supports copy operations using access points only when the source and destination buckets are in the same Amazon Web Services Region.

Alternatively, for objects accessed through Amazon S3 on Outposts, specify the ARN of the object as accessed in the format arn:aws:s3-outposts:<Region>:<account-id>:outpost/<outpost-id>/object/<key>. For example, to copy the object reports/january.pdf through outpost my-outpost owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3-outposts:us-west-2:123456789012:outpost/my-outpost/object/reports/january.pdf. The value must be URL encoded.

To copy a specific version of an object, append ?versionId=<version-id> to the value (for example, awsexamplebucket/reports/january.pdf?versionId=QUpfdndhfd8438MNFDN93jdnJFkdmqnh893). If you don't specify a version ID, Amazon S3 copies the latest version of the source object.

$sel:key:UploadPartCopy', uploadPartCopy_key - Object key for which the multipart upload was initiated.

$sel:partNumber:UploadPartCopy', uploadPartCopy_partNumber - Part number of part being copied. This is a positive integer between 1 and 10,000.

$sel:uploadId:UploadPartCopy', uploadPartCopy_uploadId - Upload ID identifying the multipart upload whose part is being copied.

Request Lenses

uploadPartCopy_copySourceIfModifiedSince :: Lens' UploadPartCopy (Maybe UTCTime)

Copies the object if it has been modified since the specified time.

uploadPartCopy_copySourceIfUnmodifiedSince :: Lens' UploadPartCopy (Maybe UTCTime)

Copies the object if it hasn't been modified since the specified time.

uploadPartCopy_copySourceRange :: Lens' UploadPartCopy (Maybe Text)

The range of bytes to copy from the source object. The range value must use the form bytes=first-last, where the first and last are the zero-based byte offsets to copy. For example, bytes=0-9 indicates that you want to copy the first 10 bytes of the source. You can copy a range only if the source object is greater than 5 MB.

uploadPartCopy_copySourceSSECustomerKeyMD5 :: Lens' UploadPartCopy (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

uploadPartCopy_copySourceIfNoneMatch :: Lens' UploadPartCopy (Maybe Text)

Copies the object if its entity tag (ETag) is different than the specified ETag.

uploadPartCopy_sSECustomerAlgorithm :: Lens' UploadPartCopy (Maybe Text)

Specifies the algorithm to use to when encrypting the object (for example, AES256).

uploadPartCopy_sSECustomerKey :: Lens' UploadPartCopy (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon S3 does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side-encryption-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

uploadPartCopy_requestPayer :: Lens' UploadPartCopy (Maybe RequestPayer)

Undocumented member.

uploadPartCopy_copySourceIfMatch :: Lens' UploadPartCopy (Maybe Text)

Copies the object if its entity tag (ETag) matches the specified tag.

uploadPartCopy_expectedSourceBucketOwner :: Lens' UploadPartCopy (Maybe Text)

The account ID of the expected source bucket owner. If the source bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

uploadPartCopy_sSECustomerKeyMD5 :: Lens' UploadPartCopy (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure that the encryption key was transmitted without error.

uploadPartCopy_copySourceSSECustomerKey :: Lens' UploadPartCopy (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

uploadPartCopy_copySourceSSECustomerAlgorithm :: Lens' UploadPartCopy (Maybe Text)

Specifies the algorithm to use when decrypting the source object (for example, AES256).

uploadPartCopy_expectedBucketOwner :: Lens' UploadPartCopy (Maybe Text)

The account ID of the expected destination bucket owner. If the destination bucket is owned by a different account, the request will fail with an HTTP 403 (Access Denied) error.

uploadPartCopy_bucket :: Lens' UploadPartCopy BucketName

The bucket name.

When using this action with an access point, you must direct requests to the access point hostname. The access point hostname takes the form AccessPointName-AccountId.s3-accesspoint.Region.amazonaws.com. When using this action with an access point through the Amazon Web Services SDKs, you provide the access point ARN in place of the bucket name. For more information about access point ARNs, see Using access points in the Amazon S3 User Guide.

When using this action with Amazon S3 on Outposts, you must direct requests to the S3 on Outposts hostname. The S3 on Outposts hostname takes the form AccessPointName-AccountId.outpostID.s3-outposts.Region.amazonaws.com. When using this action using S3 on Outposts through the Amazon Web Services SDKs, you provide the Outposts bucket ARN in place of the bucket name. For more information about S3 on Outposts ARNs, see Using S3 on Outposts in the Amazon S3 User Guide.

uploadPartCopy_copySource :: Lens' UploadPartCopy Text

Specifies the source object for the copy operation. You specify the value in one of two formats, depending on whether you want to access the source object through an access point:

For objects not accessed through an access point, specify the name of the source bucket and key of the source object, separated by a slash (/). For example, to copy the object reports/january.pdf from the bucket awsexamplebucket, use awsexamplebucket/reports/january.pdf. The value must be URL encoded.

For objects accessed through access points, specify the Amazon Resource Name (ARN) of the object as accessed through the access point, in the format arn:aws:s3:<Region>:<account-id>:accesspoint/<access-point-name>/object/<key>. For example, to copy the object reports/january.pdf through access point my-access-point owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3:us-west-2:123456789012:accesspoint/my-access-point/object/reports/january.pdf. The value must be URL encoded.

Amazon S3 supports copy operations using access points only when the source and destination buckets are in the same Amazon Web Services Region.

Alternatively, for objects accessed through Amazon S3 on Outposts, specify the ARN of the object as accessed in the format arn:aws:s3-outposts:<Region>:<account-id>:outpost/<outpost-id>/object/<key>. For example, to copy the object reports/january.pdf through outpost my-outpost owned by account 123456789012 in Region us-west-2, use the URL encoding of arn:aws:s3-outposts:us-west-2:123456789012:outpost/my-outpost/object/reports/january.pdf. The value must be URL encoded.

To copy a specific version of an object, append ?versionId=<version-id> to the value (for example, awsexamplebucket/reports/january.pdf?versionId=QUpfdndhfd8438MNFDN93jdnJFkdmqnh893). If you don't specify a version ID, Amazon S3 copies the latest version of the source object.

uploadPartCopy_key :: Lens' UploadPartCopy ObjectKey

Object key for which the multipart upload was initiated.

uploadPartCopy_partNumber :: Lens' UploadPartCopy Int

Part number of part being copied. This is a positive integer between 1 and 10,000.

uploadPartCopy_uploadId :: Lens' UploadPartCopy Text

Upload ID identifying the multipart upload whose part is being copied.

Destructuring the Response

data UploadPartCopyResponse

See: newUploadPartCopyResponse smart constructor.

Constructors

UploadPartCopyResponse'

Fields

requestCharged :: Maybe RequestCharged

copyPartResult :: Maybe CopyPartResult

Container for all response elements.

sSECustomerAlgorithm :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

bucketKeyEnabled :: Maybe Bool

Indicates whether the multipart upload uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

copySourceVersionId :: Maybe Text

The version of the source object that was copied, if you have enabled versioning on the source bucket.

sSECustomerKeyMD5 :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

sSEKMSKeyId :: Maybe (Sensitive Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

serverSideEncryption :: Maybe ServerSideEncryption

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

httpStatus :: Int

The response's http status code.

newUploadPartCopyResponse

Arguments

:: Int
$sel:httpStatus:UploadPartCopyResponse'

-> UploadPartCopyResponse

Create a value of UploadPartCopyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharged:UploadPartCopyResponse', uploadPartCopyResponse_requestCharged - Undocumented member.

$sel:copyPartResult:UploadPartCopyResponse', uploadPartCopyResponse_copyPartResult - Container for all response elements.

$sel:sSECustomerAlgorithm:UploadPartCopy', uploadPartCopyResponse_sSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

$sel:bucketKeyEnabled:UploadPartCopyResponse', uploadPartCopyResponse_bucketKeyEnabled - Indicates whether the multipart upload uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

$sel:copySourceVersionId:UploadPartCopyResponse', uploadPartCopyResponse_copySourceVersionId - The version of the source object that was copied, if you have enabled versioning on the source bucket.

$sel:sSECustomerKeyMD5:UploadPartCopy', uploadPartCopyResponse_sSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

$sel:sSEKMSKeyId:UploadPartCopyResponse', uploadPartCopyResponse_sSEKMSKeyId - If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

$sel:serverSideEncryption:UploadPartCopyResponse', uploadPartCopyResponse_serverSideEncryption - The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

$sel:httpStatus:UploadPartCopyResponse', uploadPartCopyResponse_httpStatus - The response's http status code.

Response Lenses

uploadPartCopyResponse_requestCharged :: Lens' UploadPartCopyResponse (Maybe RequestCharged)

Undocumented member.

uploadPartCopyResponse_copyPartResult :: Lens' UploadPartCopyResponse (Maybe CopyPartResult)

Container for all response elements.

uploadPartCopyResponse_sSECustomerAlgorithm :: Lens' UploadPartCopyResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

uploadPartCopyResponse_bucketKeyEnabled :: Lens' UploadPartCopyResponse (Maybe Bool)

Indicates whether the multipart upload uses an S3 Bucket Key for server-side encryption with Amazon Web Services KMS (SSE-KMS).

uploadPartCopyResponse_copySourceVersionId :: Lens' UploadPartCopyResponse (Maybe Text)

The version of the source object that was copied, if you have enabled versioning on the source bucket.

uploadPartCopyResponse_sSECustomerKeyMD5 :: Lens' UploadPartCopyResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round-trip message integrity verification of the customer-provided encryption key.

uploadPartCopyResponse_sSEKMSKeyId :: Lens' UploadPartCopyResponse (Maybe Text)

If present, specifies the ID of the Amazon Web Services Key Management Service (Amazon Web Services KMS) symmetric customer managed key that was used for the object.

uploadPartCopyResponse_serverSideEncryption :: Lens' UploadPartCopyResponse (Maybe ServerSideEncryption)

The server-side encryption algorithm used when storing this object in Amazon S3 (for example, AES256, aws:kms).

uploadPartCopyResponse_httpStatus :: Lens' UploadPartCopyResponse Int

The response's http status code.