Amazonka.S3Outposts

Description

Derived from API version 2017-07-25 of the AWS service descriptions, licensed under Apache 2.0.

Amazon S3 on Outposts provides access to S3 on Outposts operations.

Service Configuration

defaultService :: Service

API version 2017-07-25 of the Amazon S3 on Outposts SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by S3Outposts.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

There was an exception validating this data.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Access was denied for this action.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

There was a conflict with this action, and it could not be completed.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

There was an exception with the internal server.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource was not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

CreateEndpoint

data CreateEndpoint

See: newCreateEndpoint smart constructor.

Constructors

CreateEndpoint' (Maybe EndpointAccessType) (Maybe Text) Text Text Text

newCreateEndpoint

Arguments

:: Text
$sel:outpostId:CreateEndpoint'

-> Text
$sel:subnetId:CreateEndpoint'

-> Text
$sel:securityGroupId:CreateEndpoint'

-> CreateEndpoint

Create a value of CreateEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessType:CreateEndpoint', createEndpoint_accessType - The type of access for the on-premise network connectivity for the Outpost endpoint. To access the endpoint from an on-premises network, you must specify the access type and provide the customer owned IPv4 pool.

$sel:customerOwnedIpv4Pool:CreateEndpoint', createEndpoint_customerOwnedIpv4Pool - The ID of the customer-owned IPv4 pool for the endpoint. IP addresses will be allocated from this pool for the endpoint.

$sel:outpostId:CreateEndpoint', createEndpoint_outpostId - The ID of the AWS Outposts.

$sel:subnetId:CreateEndpoint', createEndpoint_subnetId - The ID of the subnet in the selected VPC. The endpoint subnet must belong to the Outpost that has the Amazon S3 on Outposts provisioned.

$sel:securityGroupId:CreateEndpoint', createEndpoint_securityGroupId - The ID of the security group to use with the endpoint.

data CreateEndpointResponse

See: newCreateEndpointResponse smart constructor.

Constructors

CreateEndpointResponse' (Maybe Text) Int

newCreateEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreateEndpointResponse'

-> CreateEndpointResponse

Create a value of CreateEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointArn:CreateEndpointResponse', createEndpointResponse_endpointArn - The Amazon Resource Name (ARN) of the endpoint.

$sel:httpStatus:CreateEndpointResponse', createEndpointResponse_httpStatus - The response's http status code.

DeleteEndpoint

data DeleteEndpoint

See: newDeleteEndpoint smart constructor.

Constructors

DeleteEndpoint' Text Text

newDeleteEndpoint

Arguments

:: Text
$sel:endpointId:DeleteEndpoint'

-> Text
$sel:outpostId:DeleteEndpoint'

-> DeleteEndpoint

Create a value of DeleteEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointId:DeleteEndpoint', deleteEndpoint_endpointId - The ID of the endpoint.

$sel:outpostId:DeleteEndpoint', deleteEndpoint_outpostId - The ID of the AWS Outposts.

data DeleteEndpointResponse

See: newDeleteEndpointResponse smart constructor.

Constructors

DeleteEndpointResponse'

newDeleteEndpointResponse :: DeleteEndpointResponse

Create a value of DeleteEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

ListEndpoints (Paginated)

data ListEndpoints

See: newListEndpoints smart constructor.

Constructors

ListEndpoints' (Maybe Text) (Maybe Natural)

newListEndpoints :: ListEndpoints

Create a value of ListEndpoints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEndpoints', listEndpoints_nextToken - The next endpoint requested in the list.

$sel:maxResults:ListEndpoints', listEndpoints_maxResults - The max number of endpoints that can be returned on the request.

data ListEndpointsResponse

See: newListEndpointsResponse smart constructor.

Constructors

ListEndpointsResponse' (Maybe Text) (Maybe [Endpoint]) Int

newListEndpointsResponse

Arguments

:: Int
$sel:httpStatus:ListEndpointsResponse'

-> ListEndpointsResponse

Create a value of ListEndpointsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEndpoints', listEndpointsResponse_nextToken - The next endpoint returned in the list.

$sel:endpoints:ListEndpointsResponse', listEndpointsResponse_endpoints - Returns an array of endpoints associated with AWS Outposts.

$sel:httpStatus:ListEndpointsResponse', listEndpointsResponse_httpStatus - The response's http status code.

Types

EndpointAccessType

newtype EndpointAccessType

Constructors

EndpointAccessType'

Fields

fromEndpointAccessType :: Text

Bundled Patterns

pattern EndpointAccessType_CustomerOwnedIp :: EndpointAccessType

pattern EndpointAccessType_Private :: EndpointAccessType

EndpointStatus

newtype EndpointStatus

Constructors

EndpointStatus'

Fields

fromEndpointStatus :: Text

Bundled Patterns

pattern EndpointStatus_Available :: EndpointStatus

pattern EndpointStatus_Deleting :: EndpointStatus

pattern EndpointStatus_Pending :: EndpointStatus

Endpoint

data Endpoint

Amazon S3 on Outposts Access Points simplify managing data access at scale for shared datasets in S3 on Outposts. S3 on Outposts uses endpoints to connect to Outposts buckets so that you can perform actions within your virtual private cloud (VPC). For more information, see Accessing S3 on Outposts using VPC only access points.

See: newEndpoint smart constructor.

Constructors

Endpoint' (Maybe POSIX) (Maybe EndpointStatus) (Maybe Text) (Maybe [NetworkInterface]) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe EndpointAccessType) (Maybe Text) (Maybe Text) (Maybe Text)

newEndpoint :: Endpoint

Create a value of Endpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Endpoint', endpoint_creationTime - The time the endpoint was created.

$sel:status:Endpoint', endpoint_status - The status of the endpoint.

$sel:vpcId:Endpoint', endpoint_vpcId - The ID of the VPC used for the endpoint.

$sel:networkInterfaces:Endpoint', endpoint_networkInterfaces - The network interface of the endpoint.

$sel:outpostsId:Endpoint', endpoint_outpostsId - The ID of the AWS Outposts.

$sel:subnetId:Endpoint', endpoint_subnetId - The ID of the subnet used for the endpoint.

$sel:securityGroupId:Endpoint', endpoint_securityGroupId - The ID of the security group used for the endpoint.

$sel:accessType:Endpoint', endpoint_accessType -

$sel:customerOwnedIpv4Pool:Endpoint', endpoint_customerOwnedIpv4Pool - The ID of the customer-owned IPv4 pool used for the endpoint.

$sel:cidrBlock:Endpoint', endpoint_cidrBlock - The VPC CIDR committed by this endpoint.

$sel:endpointArn:Endpoint', endpoint_endpointArn - The Amazon Resource Name (ARN) of the endpoint.

NetworkInterface

data NetworkInterface

The container for the network interface.

See: newNetworkInterface smart constructor.

Constructors

NetworkInterface' (Maybe Text)

newNetworkInterface :: NetworkInterface

Create a value of NetworkInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkInterfaceId:NetworkInterface', networkInterface_networkInterfaceId - The ID for the network interface.