Amazonka.S3Outposts.Lens

Description

Operations

CreateEndpoint

createEndpoint_accessType :: Lens' CreateEndpoint (Maybe EndpointAccessType)

The type of access for the on-premise network connectivity for the Outpost endpoint. To access the endpoint from an on-premises network, you must specify the access type and provide the customer owned IPv4 pool.

createEndpoint_customerOwnedIpv4Pool :: Lens' CreateEndpoint (Maybe Text)

The ID of the customer-owned IPv4 pool for the endpoint. IP addresses will be allocated from this pool for the endpoint.

createEndpoint_outpostId :: Lens' CreateEndpoint Text

The ID of the AWS Outposts.

createEndpoint_subnetId :: Lens' CreateEndpoint Text

The ID of the subnet in the selected VPC. The endpoint subnet must belong to the Outpost that has the Amazon S3 on Outposts provisioned.

createEndpoint_securityGroupId :: Lens' CreateEndpoint Text

The ID of the security group to use with the endpoint.

createEndpointResponse_endpointArn :: Lens' CreateEndpointResponse (Maybe Text)

The Amazon Resource Name (ARN) of the endpoint.

createEndpointResponse_httpStatus :: Lens' CreateEndpointResponse Int

The response's http status code.

DeleteEndpoint

deleteEndpoint_endpointId :: Lens' DeleteEndpoint Text

The ID of the endpoint.

deleteEndpoint_outpostId :: Lens' DeleteEndpoint Text

The ID of the AWS Outposts.

ListEndpoints

listEndpoints_nextToken :: Lens' ListEndpoints (Maybe Text)

The next endpoint requested in the list.

listEndpoints_maxResults :: Lens' ListEndpoints (Maybe Natural)

The max number of endpoints that can be returned on the request.

listEndpointsResponse_nextToken :: Lens' ListEndpointsResponse (Maybe Text)

The next endpoint returned in the list.

listEndpointsResponse_endpoints :: Lens' ListEndpointsResponse (Maybe [Endpoint])

Returns an array of endpoints associated with AWS Outposts.

listEndpointsResponse_httpStatus :: Lens' ListEndpointsResponse Int

The response's http status code.

Types

Endpoint

endpoint_creationTime :: Lens' Endpoint (Maybe UTCTime)

The time the endpoint was created.

endpoint_status :: Lens' Endpoint (Maybe EndpointStatus)

The status of the endpoint.

endpoint_vpcId :: Lens' Endpoint (Maybe Text)

The ID of the VPC used for the endpoint.

endpoint_networkInterfaces :: Lens' Endpoint (Maybe [NetworkInterface])

The network interface of the endpoint.

endpoint_outpostsId :: Lens' Endpoint (Maybe Text)

The ID of the AWS Outposts.

endpoint_subnetId :: Lens' Endpoint (Maybe Text)

The ID of the subnet used for the endpoint.

endpoint_securityGroupId :: Lens' Endpoint (Maybe Text)

The ID of the security group used for the endpoint.

endpoint_accessType :: Lens' Endpoint (Maybe EndpointAccessType)

endpoint_customerOwnedIpv4Pool :: Lens' Endpoint (Maybe Text)

The ID of the customer-owned IPv4 pool used for the endpoint.

endpoint_cidrBlock :: Lens' Endpoint (Maybe Text)

The VPC CIDR committed by this endpoint.

endpoint_endpointArn :: Lens' Endpoint (Maybe Text)

The Amazon Resource Name (ARN) of the endpoint.

NetworkInterface

networkInterface_networkInterfaceId :: Lens' NetworkInterface (Maybe Text)

The ID for the network interface.