Amazonka.SES.DeleteReceiptFilter

Description

Deletes the specified IP address filter.

For information about managing IP address filters, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data DeleteReceiptFilter

Represents a request to delete an IP address filter. You use IP address filters when you receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

See: newDeleteReceiptFilter smart constructor.

Constructors

DeleteReceiptFilter'

Fields

filterName :: Text

The name of the IP address filter to delete.

newDeleteReceiptFilter

Arguments

:: Text
$sel:filterName:DeleteReceiptFilter'

-> DeleteReceiptFilter

Create a value of DeleteReceiptFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filterName:DeleteReceiptFilter', deleteReceiptFilter_filterName - The name of the IP address filter to delete.

Request Lenses

deleteReceiptFilter_filterName :: Lens' DeleteReceiptFilter Text

The name of the IP address filter to delete.

Destructuring the Response

data DeleteReceiptFilterResponse

An empty element returned on a successful request.

See: newDeleteReceiptFilterResponse smart constructor.

Constructors

DeleteReceiptFilterResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteReceiptFilterResponse

Arguments

:: Int
$sel:httpStatus:DeleteReceiptFilterResponse'

-> DeleteReceiptFilterResponse

Create a value of DeleteReceiptFilterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteReceiptFilterResponse', deleteReceiptFilterResponse_httpStatus - The response's http status code.

Response Lenses

deleteReceiptFilterResponse_httpStatus :: Lens' DeleteReceiptFilterResponse Int

The response's http status code.