Amazonka.SES.ListConfigurationSets

Description

Provides a list of the configuration sets associated with your Amazon SES account in the current AWS Region. For information about using configuration sets, see Monitoring Your Amazon SES Sending Activity in the Amazon SES Developer Guide.

You can execute this operation no more than once per second. This operation will return up to 1,000 configuration sets each time it is run. If your Amazon SES account has more than 1,000 configuration sets, this operation will also return a NextToken element. You can then execute the ListConfigurationSets operation again, passing the NextToken parameter and the value of the NextToken element to retrieve additional results.

This operation returns paginated results.

Creating a Request

data ListConfigurationSets

Represents a request to list the configuration sets associated with your AWS account. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

See: newListConfigurationSets smart constructor.

Constructors

ListConfigurationSets'

Fields

nextToken :: Maybe Text

A token returned from a previous call to ListConfigurationSets to indicate the position of the configuration set in the configuration set list.

maxItems :: Maybe Int

The number of configuration sets to return.

newListConfigurationSets :: ListConfigurationSets

Create a value of ListConfigurationSets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListConfigurationSets', listConfigurationSets_nextToken - A token returned from a previous call to ListConfigurationSets to indicate the position of the configuration set in the configuration set list.

$sel:maxItems:ListConfigurationSets', listConfigurationSets_maxItems - The number of configuration sets to return.

Request Lenses

listConfigurationSets_nextToken :: Lens' ListConfigurationSets (Maybe Text)

A token returned from a previous call to ListConfigurationSets to indicate the position of the configuration set in the configuration set list.

listConfigurationSets_maxItems :: Lens' ListConfigurationSets (Maybe Int)

The number of configuration sets to return.

Destructuring the Response

data ListConfigurationSetsResponse

A list of configuration sets associated with your AWS account. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

See: newListConfigurationSetsResponse smart constructor.

Constructors

ListConfigurationSetsResponse'

Fields

configurationSets :: Maybe [ConfigurationSet]

A list of configuration sets.

nextToken :: Maybe Text

A token indicating that there are additional configuration sets available to be listed. Pass this token to successive calls of ListConfigurationSets.

httpStatus :: Int

The response's http status code.

newListConfigurationSetsResponse

Arguments

:: Int
$sel:httpStatus:ListConfigurationSetsResponse'

-> ListConfigurationSetsResponse

Create a value of ListConfigurationSetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSets:ListConfigurationSetsResponse', listConfigurationSetsResponse_configurationSets - A list of configuration sets.

$sel:nextToken:ListConfigurationSets', listConfigurationSetsResponse_nextToken - A token indicating that there are additional configuration sets available to be listed. Pass this token to successive calls of ListConfigurationSets.

$sel:httpStatus:ListConfigurationSetsResponse', listConfigurationSetsResponse_httpStatus - The response's http status code.

Response Lenses

listConfigurationSetsResponse_configurationSets :: Lens' ListConfigurationSetsResponse (Maybe [ConfigurationSet])

A list of configuration sets.

listConfigurationSetsResponse_nextToken :: Lens' ListConfigurationSetsResponse (Maybe Text)

A token indicating that there are additional configuration sets available to be listed. Pass this token to successive calls of ListConfigurationSets.

listConfigurationSetsResponse_httpStatus :: Lens' ListConfigurationSetsResponse Int

The response's http status code.