Amazonka.SES.PutIdentityPolicy

Description

Adds or updates a sending authorization policy for the specified identity (an email address or a domain).

This API is for the identity owner only. If you have not verified the identity, this API will return an error.

Sending authorization is a feature that enables an identity owner to authorize other senders to use its identities. For information about using sending authorization, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data PutIdentityPolicy

Represents a request to add or update a sending authorization policy for an identity. Sending authorization is an Amazon SES feature that enables you to authorize other senders to use your identities. For information, see the Amazon SES Developer Guide.

See: newPutIdentityPolicy smart constructor.

Constructors

PutIdentityPolicy'

Fields

identity :: Text

The identity that the policy will apply to. You can specify an identity by using its name or by using its Amazon Resource Name (ARN). Examples: user@example.com, example.com, arn:aws:ses:us-east-1:123456789012:identity/example.com.

To successfully call this API, you must own the identity.

policyName :: Text

The name of the policy.

The policy name cannot exceed 64 characters and can only include alphanumeric characters, dashes, and underscores.

policy :: Text

The text of the policy in JSON format. The policy cannot exceed 4 KB.

For information about the syntax of sending authorization policies, see the Amazon SES Developer Guide.

newPutIdentityPolicy

Arguments

:: Text
$sel:identity:PutIdentityPolicy'

-> Text
$sel:policyName:PutIdentityPolicy'

-> Text
$sel:policy:PutIdentityPolicy'

-> PutIdentityPolicy

Create a value of PutIdentityPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identity:PutIdentityPolicy', putIdentityPolicy_identity - The identity that the policy will apply to. You can specify an identity by using its name or by using its Amazon Resource Name (ARN). Examples: user@example.com, example.com, arn:aws:ses:us-east-1:123456789012:identity/example.com.

To successfully call this API, you must own the identity.

$sel:policyName:PutIdentityPolicy', putIdentityPolicy_policyName - The name of the policy.

The policy name cannot exceed 64 characters and can only include alphanumeric characters, dashes, and underscores.

$sel:policy:PutIdentityPolicy', putIdentityPolicy_policy - The text of the policy in JSON format. The policy cannot exceed 4 KB.

For information about the syntax of sending authorization policies, see the Amazon SES Developer Guide.

Request Lenses

putIdentityPolicy_identity :: Lens' PutIdentityPolicy Text

The identity that the policy will apply to. You can specify an identity by using its name or by using its Amazon Resource Name (ARN). Examples: user@example.com, example.com, arn:aws:ses:us-east-1:123456789012:identity/example.com.

To successfully call this API, you must own the identity.

putIdentityPolicy_policyName :: Lens' PutIdentityPolicy Text

The name of the policy.

The policy name cannot exceed 64 characters and can only include alphanumeric characters, dashes, and underscores.

putIdentityPolicy_policy :: Lens' PutIdentityPolicy Text

The text of the policy in JSON format. The policy cannot exceed 4 KB.

For information about the syntax of sending authorization policies, see the Amazon SES Developer Guide.

Destructuring the Response

data PutIdentityPolicyResponse

An empty element returned on a successful request.

See: newPutIdentityPolicyResponse smart constructor.

Constructors

PutIdentityPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutIdentityPolicyResponse

Arguments

:: Int
$sel:httpStatus:PutIdentityPolicyResponse'

-> PutIdentityPolicyResponse

Create a value of PutIdentityPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutIdentityPolicyResponse', putIdentityPolicyResponse_httpStatus - The response's http status code.

Response Lenses

putIdentityPolicyResponse_httpStatus :: Lens' PutIdentityPolicyResponse Int

The response's http status code.