Amazonka.SES.SendBulkTemplatedEmail

Description

Composes an email message to multiple destinations. The message body is created using an email template.

In order to send email using the SendBulkTemplatedEmail operation, your call to the API must meet the following requirements:

The call must refer to an existing email template. You can create email templates using the CreateTemplate operation.

The message must be sent from a verified email address or domain.

If your account is still in the Amazon SES sandbox, you may only send to verified addresses or domains, or to email addresses associated with the Amazon SES Mailbox Simulator. For more information, see Verifying Email Addresses and Domains in the Amazon SES Developer Guide.

The maximum message size is 10 MB.

Each Destination parameter must include at least one recipient email address. The recipient address can be a To: address, a CC: address, or a BCC: address. If a recipient email address is invalid (that is, it is not in the format UserName@[SubDomain.]Domain.TopLevelDomain), the entire message will be rejected, even if the message contains other recipients that are valid.

The message may not include more than 50 recipients, across the To:, CC: and BCC: fields. If you need to send an email message to a larger audience, you can divide your recipient list into groups of 50 or fewer, and then call the SendBulkTemplatedEmail operation several times to send the message to each group.

The number of destinations you can contact in a single call to the API may be limited by your account's maximum sending rate.

Creating a Request

data SendBulkTemplatedEmail

Represents a request to send a templated email to multiple destinations using Amazon SES. For more information, see the Amazon SES Developer Guide.

See: newSendBulkTemplatedEmail smart constructor.

Constructors

SendBulkTemplatedEmail'

Fields

returnPath :: Maybe Text

The email address that bounces and complaints will be forwarded to when feedback forwarding is enabled. If the message cannot be delivered to the recipient, then an error message will be returned from the recipient's ISP; this message will then be forwarded to the email address specified by the ReturnPath parameter. The ReturnPath parameter is never overwritten. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES.

configurationSetName :: Maybe Text

The name of the configuration set to use when you send an email using SendBulkTemplatedEmail.

sourceArn :: Maybe Text

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to send for the email address specified in the Source parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to send from user@example.com, then you would specify the SourceArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the Source to be user@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

defaultTags :: Maybe [MessageTag]

A list of tags, in the form of name/value pairs, to apply to an email that you send to a destination using SendBulkTemplatedEmail.

returnPathArn :: Maybe Text

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the ReturnPath parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use feedback@example.com, then you would specify the ReturnPathArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the ReturnPath to be feedback@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

templateArn :: Maybe Text

The ARN of the template to use when sending this email.

defaultTemplateData :: Maybe Text

A list of replacement values to apply to the template when replacement data is not specified in a Destination object. These values act as a default or fallback option when no other data is available.

The template data is a JSON object, typically consisting of key-value pairs in which the keys correspond to replacement tags in the email template.

replyToAddresses :: Maybe [Text]

The reply-to email address(es) for the message. If the recipient replies to the message, each reply-to address will receive the reply.

source :: Text

The email address that is sending the email. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES. For information about verifying identities, see the Amazon SES Developer Guide.

If you are sending on behalf of another user and have been permitted to do so by a sending authorization policy, then you must also specify the SourceArn parameter. For more information about sending authorization, see the Amazon SES Developer Guide.

Amazon SES does not support the SMTPUTF8 extension, as described in RFC6531. For this reason, the local part of a source email address (the part of the email address that precedes the @ sign) may only contain 7-bit ASCII characters. If the domain part of an address (the part after the @ sign) contains non-ASCII characters, they must be encoded using Punycode, as described in RFC3492. The sender name (also known as the friendly name) may contain non-ASCII characters. These characters must be encoded using MIME encoded-word syntax, as described in RFC 2047. MIME encoded-word syntax uses the following form: =?charset?encoding?encoded-text?=.

template :: Text

The template to use when sending this email.

destinations :: [BulkEmailDestination]

One or more Destination objects. All of the recipients in a Destination will receive the same version of the email. You can specify up to 50 Destination objects within a Destinations array.

newSendBulkTemplatedEmail

Arguments

:: Text
$sel:source:SendBulkTemplatedEmail'

-> Text
$sel:template:SendBulkTemplatedEmail'

-> SendBulkTemplatedEmail

Create a value of SendBulkTemplatedEmail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:returnPath:SendBulkTemplatedEmail', sendBulkTemplatedEmail_returnPath - The email address that bounces and complaints will be forwarded to when feedback forwarding is enabled. If the message cannot be delivered to the recipient, then an error message will be returned from the recipient's ISP; this message will then be forwarded to the email address specified by the ReturnPath parameter. The ReturnPath parameter is never overwritten. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES.

$sel:configurationSetName:SendBulkTemplatedEmail', sendBulkTemplatedEmail_configurationSetName - The name of the configuration set to use when you send an email using SendBulkTemplatedEmail.

$sel:sourceArn:SendBulkTemplatedEmail', sendBulkTemplatedEmail_sourceArn - This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to send for the email address specified in the Source parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to send from user@example.com, then you would specify the SourceArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the Source to be user@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

$sel:defaultTags:SendBulkTemplatedEmail', sendBulkTemplatedEmail_defaultTags - A list of tags, in the form of name/value pairs, to apply to an email that you send to a destination using SendBulkTemplatedEmail.

$sel:returnPathArn:SendBulkTemplatedEmail', sendBulkTemplatedEmail_returnPathArn - This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the ReturnPath parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use feedback@example.com, then you would specify the ReturnPathArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the ReturnPath to be feedback@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

$sel:templateArn:SendBulkTemplatedEmail', sendBulkTemplatedEmail_templateArn - The ARN of the template to use when sending this email.

$sel:defaultTemplateData:SendBulkTemplatedEmail', sendBulkTemplatedEmail_defaultTemplateData - A list of replacement values to apply to the template when replacement data is not specified in a Destination object. These values act as a default or fallback option when no other data is available.

The template data is a JSON object, typically consisting of key-value pairs in which the keys correspond to replacement tags in the email template.

$sel:replyToAddresses:SendBulkTemplatedEmail', sendBulkTemplatedEmail_replyToAddresses - The reply-to email address(es) for the message. If the recipient replies to the message, each reply-to address will receive the reply.

$sel:source:SendBulkTemplatedEmail', sendBulkTemplatedEmail_source - The email address that is sending the email. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES. For information about verifying identities, see the Amazon SES Developer Guide.

If you are sending on behalf of another user and have been permitted to do so by a sending authorization policy, then you must also specify the SourceArn parameter. For more information about sending authorization, see the Amazon SES Developer Guide.

Amazon SES does not support the SMTPUTF8 extension, as described in RFC6531. For this reason, the local part of a source email address (the part of the email address that precedes the @ sign) may only contain 7-bit ASCII characters. If the domain part of an address (the part after the @ sign) contains non-ASCII characters, they must be encoded using Punycode, as described in RFC3492. The sender name (also known as the friendly name) may contain non-ASCII characters. These characters must be encoded using MIME encoded-word syntax, as described in RFC 2047. MIME encoded-word syntax uses the following form: =?charset?encoding?encoded-text?=.

$sel:template:SendBulkTemplatedEmail', sendBulkTemplatedEmail_template - The template to use when sending this email.

$sel:destinations:SendBulkTemplatedEmail', sendBulkTemplatedEmail_destinations - One or more Destination objects. All of the recipients in a Destination will receive the same version of the email. You can specify up to 50 Destination objects within a Destinations array.

Request Lenses

sendBulkTemplatedEmail_returnPath :: Lens' SendBulkTemplatedEmail (Maybe Text)

The email address that bounces and complaints will be forwarded to when feedback forwarding is enabled. If the message cannot be delivered to the recipient, then an error message will be returned from the recipient's ISP; this message will then be forwarded to the email address specified by the ReturnPath parameter. The ReturnPath parameter is never overwritten. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES.

sendBulkTemplatedEmail_configurationSetName :: Lens' SendBulkTemplatedEmail (Maybe Text)

The name of the configuration set to use when you send an email using SendBulkTemplatedEmail.

sendBulkTemplatedEmail_sourceArn :: Lens' SendBulkTemplatedEmail (Maybe Text)

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to send for the email address specified in the Source parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to send from user@example.com, then you would specify the SourceArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the Source to be user@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

sendBulkTemplatedEmail_defaultTags :: Lens' SendBulkTemplatedEmail (Maybe [MessageTag])

A list of tags, in the form of name/value pairs, to apply to an email that you send to a destination using SendBulkTemplatedEmail.

sendBulkTemplatedEmail_returnPathArn :: Lens' SendBulkTemplatedEmail (Maybe Text)

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the ReturnPath parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use feedback@example.com, then you would specify the ReturnPathArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the ReturnPath to be feedback@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

sendBulkTemplatedEmail_templateArn :: Lens' SendBulkTemplatedEmail (Maybe Text)

The ARN of the template to use when sending this email.

sendBulkTemplatedEmail_defaultTemplateData :: Lens' SendBulkTemplatedEmail (Maybe Text)

A list of replacement values to apply to the template when replacement data is not specified in a Destination object. These values act as a default or fallback option when no other data is available.

The template data is a JSON object, typically consisting of key-value pairs in which the keys correspond to replacement tags in the email template.

sendBulkTemplatedEmail_replyToAddresses :: Lens' SendBulkTemplatedEmail (Maybe [Text])

The reply-to email address(es) for the message. If the recipient replies to the message, each reply-to address will receive the reply.

sendBulkTemplatedEmail_source :: Lens' SendBulkTemplatedEmail Text

The email address that is sending the email. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES. For information about verifying identities, see the Amazon SES Developer Guide.

If you are sending on behalf of another user and have been permitted to do so by a sending authorization policy, then you must also specify the SourceArn parameter. For more information about sending authorization, see the Amazon SES Developer Guide.

Amazon SES does not support the SMTPUTF8 extension, as described in RFC6531. For this reason, the local part of a source email address (the part of the email address that precedes the @ sign) may only contain 7-bit ASCII characters. If the domain part of an address (the part after the @ sign) contains non-ASCII characters, they must be encoded using Punycode, as described in RFC3492. The sender name (also known as the friendly name) may contain non-ASCII characters. These characters must be encoded using MIME encoded-word syntax, as described in RFC 2047. MIME encoded-word syntax uses the following form: =?charset?encoding?encoded-text?=.

sendBulkTemplatedEmail_template :: Lens' SendBulkTemplatedEmail Text

The template to use when sending this email.

sendBulkTemplatedEmail_destinations :: Lens' SendBulkTemplatedEmail [BulkEmailDestination]

One or more Destination objects. All of the recipients in a Destination will receive the same version of the email. You can specify up to 50 Destination objects within a Destinations array.

Destructuring the Response

data SendBulkTemplatedEmailResponse

See: newSendBulkTemplatedEmailResponse smart constructor.

Constructors

SendBulkTemplatedEmailResponse'

Fields

httpStatus :: Int

The response's http status code.

status :: [BulkEmailDestinationStatus]

The unique message identifier returned from the SendBulkTemplatedEmail action.

newSendBulkTemplatedEmailResponse

Arguments

:: Int
$sel:httpStatus:SendBulkTemplatedEmailResponse'

-> SendBulkTemplatedEmailResponse

Create a value of SendBulkTemplatedEmailResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SendBulkTemplatedEmailResponse', sendBulkTemplatedEmailResponse_httpStatus - The response's http status code.

$sel:status:SendBulkTemplatedEmailResponse', sendBulkTemplatedEmailResponse_status - The unique message identifier returned from the SendBulkTemplatedEmail action.

Response Lenses

sendBulkTemplatedEmailResponse_httpStatus :: Lens' SendBulkTemplatedEmailResponse Int

The response's http status code.

sendBulkTemplatedEmailResponse_status :: Lens' SendBulkTemplatedEmailResponse [BulkEmailDestinationStatus]

The unique message identifier returned from the SendBulkTemplatedEmail action.