Amazonka.SES.SendTemplatedEmail

Description

Composes an email message using an email template and immediately queues it for sending.

In order to send email using the SendTemplatedEmail operation, your call to the API must meet the following requirements:

The call must refer to an existing email template. You can create email templates using the CreateTemplate operation.

The message must be sent from a verified email address or domain.

If your account is still in the Amazon SES sandbox, you may only send to verified addresses or domains, or to email addresses associated with the Amazon SES Mailbox Simulator. For more information, see Verifying Email Addresses and Domains in the Amazon SES Developer Guide.

The maximum message size is 10 MB.

Calls to the SendTemplatedEmail operation may only include one Destination parameter. A destination is a set of recipients who will receive the same version of the email. The Destination parameter can include up to 50 recipients, across the To:, CC: and BCC: fields.

The Destination parameter must include at least one recipient email address. The recipient address can be a To: address, a CC: address, or a BCC: address. If a recipient email address is invalid (that is, it is not in the format UserName@[SubDomain.]Domain.TopLevelDomain), the entire message will be rejected, even if the message contains other recipients that are valid.

If your call to the SendTemplatedEmail operation includes all of the required parameters, Amazon SES accepts it and returns a Message ID. However, if Amazon SES can't render the email because the template contains errors, it doesn't send the email. Additionally, because it already accepted the message, Amazon SES doesn't return a message stating that it was unable to send the email.

For these reasons, we highly recommend that you set up Amazon SES to send you notifications when Rendering Failure events occur. For more information, see Sending Personalized Email Using the Amazon SES API in the Amazon Simple Email Service Developer Guide.

Creating a Request

data SendTemplatedEmail

Represents a request to send a templated email using Amazon SES. For more information, see the Amazon SES Developer Guide.

See: newSendTemplatedEmail smart constructor.

Constructors

SendTemplatedEmail'

Fields

returnPath :: Maybe Text

The email address that bounces and complaints will be forwarded to when feedback forwarding is enabled. If the message cannot be delivered to the recipient, then an error message will be returned from the recipient's ISP; this message will then be forwarded to the email address specified by the ReturnPath parameter. The ReturnPath parameter is never overwritten. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES.

configurationSetName :: Maybe Text

The name of the configuration set to use when you send an email using SendTemplatedEmail.

sourceArn :: Maybe Text

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to send for the email address specified in the Source parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to send from user@example.com, then you would specify the SourceArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the Source to be user@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

returnPathArn :: Maybe Text

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the ReturnPath parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use feedback@example.com, then you would specify the ReturnPathArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the ReturnPath to be feedback@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

templateArn :: Maybe Text

The ARN of the template to use when sending this email.

tags :: Maybe [MessageTag]

A list of tags, in the form of name/value pairs, to apply to an email that you send using SendTemplatedEmail. Tags correspond to characteristics of the email that you define, so that you can publish email sending events.

replyToAddresses :: Maybe [Text]

The reply-to email address(es) for the message. If the recipient replies to the message, each reply-to address will receive the reply.

source :: Text

The email address that is sending the email. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES. For information about verifying identities, see the Amazon SES Developer Guide.

If you are sending on behalf of another user and have been permitted to do so by a sending authorization policy, then you must also specify the SourceArn parameter. For more information about sending authorization, see the Amazon SES Developer Guide.

Amazon SES does not support the SMTPUTF8 extension, as described in RFC6531. For this reason, the local part of a source email address (the part of the email address that precedes the @ sign) may only contain 7-bit ASCII characters. If the domain part of an address (the part after the @ sign) contains non-ASCII characters, they must be encoded using Punycode, as described in RFC3492. The sender name (also known as the friendly name) may contain non-ASCII characters. These characters must be encoded using MIME encoded-word syntax, as described inRFC 2047. MIME encoded-word syntax uses the following form: =?charset?encoding?encoded-text?=.

destination :: Destination

The destination for this email, composed of To:, CC:, and BCC: fields. A Destination can include up to 50 recipients across these three fields.

template :: Text

The template to use when sending this email.

templateData :: Text

A list of replacement values to apply to the template. This parameter is a JSON object, typically consisting of key-value pairs in which the keys correspond to replacement tags in the email template.

newSendTemplatedEmail

Arguments

:: Text
$sel:source:SendTemplatedEmail'

-> Destination
$sel:destination:SendTemplatedEmail'

-> Text
$sel:template:SendTemplatedEmail'

-> Text
$sel:templateData:SendTemplatedEmail'

-> SendTemplatedEmail

Create a value of SendTemplatedEmail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:returnPath:SendTemplatedEmail', sendTemplatedEmail_returnPath - The email address that bounces and complaints will be forwarded to when feedback forwarding is enabled. If the message cannot be delivered to the recipient, then an error message will be returned from the recipient's ISP; this message will then be forwarded to the email address specified by the ReturnPath parameter. The ReturnPath parameter is never overwritten. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES.

$sel:configurationSetName:SendTemplatedEmail', sendTemplatedEmail_configurationSetName - The name of the configuration set to use when you send an email using SendTemplatedEmail.

$sel:sourceArn:SendTemplatedEmail', sendTemplatedEmail_sourceArn - This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to send for the email address specified in the Source parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to send from user@example.com, then you would specify the SourceArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the Source to be user@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

$sel:returnPathArn:SendTemplatedEmail', sendTemplatedEmail_returnPathArn - This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the ReturnPath parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use feedback@example.com, then you would specify the ReturnPathArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the ReturnPath to be feedback@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

$sel:templateArn:SendTemplatedEmail', sendTemplatedEmail_templateArn - The ARN of the template to use when sending this email.

$sel:tags:SendTemplatedEmail', sendTemplatedEmail_tags - A list of tags, in the form of name/value pairs, to apply to an email that you send using SendTemplatedEmail. Tags correspond to characteristics of the email that you define, so that you can publish email sending events.

$sel:replyToAddresses:SendTemplatedEmail', sendTemplatedEmail_replyToAddresses - The reply-to email address(es) for the message. If the recipient replies to the message, each reply-to address will receive the reply.

$sel:source:SendTemplatedEmail', sendTemplatedEmail_source - The email address that is sending the email. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES. For information about verifying identities, see the Amazon SES Developer Guide.

If you are sending on behalf of another user and have been permitted to do so by a sending authorization policy, then you must also specify the SourceArn parameter. For more information about sending authorization, see the Amazon SES Developer Guide.

Amazon SES does not support the SMTPUTF8 extension, as described in RFC6531. For this reason, the local part of a source email address (the part of the email address that precedes the @ sign) may only contain 7-bit ASCII characters. If the domain part of an address (the part after the @ sign) contains non-ASCII characters, they must be encoded using Punycode, as described in RFC3492. The sender name (also known as the friendly name) may contain non-ASCII characters. These characters must be encoded using MIME encoded-word syntax, as described inRFC 2047. MIME encoded-word syntax uses the following form: =?charset?encoding?encoded-text?=.

$sel:destination:SendTemplatedEmail', sendTemplatedEmail_destination - The destination for this email, composed of To:, CC:, and BCC: fields. A Destination can include up to 50 recipients across these three fields.

$sel:template:SendTemplatedEmail', sendTemplatedEmail_template - The template to use when sending this email.

$sel:templateData:SendTemplatedEmail', sendTemplatedEmail_templateData - A list of replacement values to apply to the template. This parameter is a JSON object, typically consisting of key-value pairs in which the keys correspond to replacement tags in the email template.

Request Lenses

sendTemplatedEmail_returnPath :: Lens' SendTemplatedEmail (Maybe Text)

The email address that bounces and complaints will be forwarded to when feedback forwarding is enabled. If the message cannot be delivered to the recipient, then an error message will be returned from the recipient's ISP; this message will then be forwarded to the email address specified by the ReturnPath parameter. The ReturnPath parameter is never overwritten. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES.

sendTemplatedEmail_configurationSetName :: Lens' SendTemplatedEmail (Maybe Text)

The name of the configuration set to use when you send an email using SendTemplatedEmail.

sendTemplatedEmail_sourceArn :: Lens' SendTemplatedEmail (Maybe Text)

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to send for the email address specified in the Source parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to send from user@example.com, then you would specify the SourceArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the Source to be user@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

sendTemplatedEmail_returnPathArn :: Lens' SendTemplatedEmail (Maybe Text)

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the ReturnPath parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use feedback@example.com, then you would specify the ReturnPathArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the ReturnPath to be feedback@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

sendTemplatedEmail_templateArn :: Lens' SendTemplatedEmail (Maybe Text)

The ARN of the template to use when sending this email.

sendTemplatedEmail_tags :: Lens' SendTemplatedEmail (Maybe [MessageTag])

A list of tags, in the form of name/value pairs, to apply to an email that you send using SendTemplatedEmail. Tags correspond to characteristics of the email that you define, so that you can publish email sending events.

sendTemplatedEmail_replyToAddresses :: Lens' SendTemplatedEmail (Maybe [Text])

The reply-to email address(es) for the message. If the recipient replies to the message, each reply-to address will receive the reply.

sendTemplatedEmail_source :: Lens' SendTemplatedEmail Text

The email address that is sending the email. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES. For information about verifying identities, see the Amazon SES Developer Guide.

If you are sending on behalf of another user and have been permitted to do so by a sending authorization policy, then you must also specify the SourceArn parameter. For more information about sending authorization, see the Amazon SES Developer Guide.

Amazon SES does not support the SMTPUTF8 extension, as described in RFC6531. For this reason, the local part of a source email address (the part of the email address that precedes the @ sign) may only contain 7-bit ASCII characters. If the domain part of an address (the part after the @ sign) contains non-ASCII characters, they must be encoded using Punycode, as described in RFC3492. The sender name (also known as the friendly name) may contain non-ASCII characters. These characters must be encoded using MIME encoded-word syntax, as described inRFC 2047. MIME encoded-word syntax uses the following form: =?charset?encoding?encoded-text?=.

sendTemplatedEmail_destination :: Lens' SendTemplatedEmail Destination

The destination for this email, composed of To:, CC:, and BCC: fields. A Destination can include up to 50 recipients across these three fields.

sendTemplatedEmail_template :: Lens' SendTemplatedEmail Text

The template to use when sending this email.

sendTemplatedEmail_templateData :: Lens' SendTemplatedEmail Text

A list of replacement values to apply to the template. This parameter is a JSON object, typically consisting of key-value pairs in which the keys correspond to replacement tags in the email template.

Destructuring the Response

data SendTemplatedEmailResponse

See: newSendTemplatedEmailResponse smart constructor.

Constructors

SendTemplatedEmailResponse'

Fields

httpStatus :: Int

The response's http status code.

messageId :: Text

The unique message identifier returned from the SendTemplatedEmail action.

newSendTemplatedEmailResponse

Arguments

:: Int
$sel:httpStatus:SendTemplatedEmailResponse'

-> Text
$sel:messageId:SendTemplatedEmailResponse'

-> SendTemplatedEmailResponse

Create a value of SendTemplatedEmailResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SendTemplatedEmailResponse', sendTemplatedEmailResponse_httpStatus - The response's http status code.

$sel:messageId:SendTemplatedEmailResponse', sendTemplatedEmailResponse_messageId - The unique message identifier returned from the SendTemplatedEmail action.

Response Lenses

sendTemplatedEmailResponse_httpStatus :: Lens' SendTemplatedEmailResponse Int

The response's http status code.

sendTemplatedEmailResponse_messageId :: Lens' SendTemplatedEmailResponse Text

The unique message identifier returned from the SendTemplatedEmail action.