Amazonka.SES.SetIdentityMailFromDomain

Description

Enables or disables the custom MAIL FROM domain setup for a verified identity (an email address or a domain).

To send emails using the specified MAIL FROM domain, you must add an MX record to your MAIL FROM domain's DNS settings. If you want your emails to pass Sender Policy Framework (SPF) checks, you must also add or update an SPF record. For more information, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data SetIdentityMailFromDomain

Represents a request to enable or disable the Amazon SES custom MAIL FROM domain setup for a verified identity. For information about using a custom MAIL FROM domain, see the Amazon SES Developer Guide.

See: newSetIdentityMailFromDomain smart constructor.

Constructors

SetIdentityMailFromDomain'

Fields

mailFromDomain :: Maybe Text

The custom MAIL FROM domain that you want the verified identity to use. The MAIL FROM domain must 1) be a subdomain of the verified identity, 2) not be used in a "From" address if the MAIL FROM domain is the destination of email feedback forwarding (for more information, see the Amazon SES Developer Guide), and 3) not be used to receive emails. A value of null disables the custom MAIL FROM setting for the identity.

behaviorOnMXFailure :: Maybe BehaviorOnMXFailure

The action that you want Amazon SES to take if it cannot successfully read the required MX record when you send an email. If you choose UseDefaultValue, Amazon SES will use amazonses.com (or a subdomain of that) as the MAIL FROM domain. If you choose RejectMessage, Amazon SES will return a MailFromDomainNotVerified error and not send the email.

The action specified in BehaviorOnMXFailure is taken when the custom MAIL FROM domain setup is in the Pending, Failed, and TemporaryFailure states.

identity :: Text

The verified identity for which you want to enable or disable the specified custom MAIL FROM domain.

newSetIdentityMailFromDomain

Arguments

:: Text
$sel:identity:SetIdentityMailFromDomain'

-> SetIdentityMailFromDomain

Create a value of SetIdentityMailFromDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mailFromDomain:SetIdentityMailFromDomain', setIdentityMailFromDomain_mailFromDomain - The custom MAIL FROM domain that you want the verified identity to use. The MAIL FROM domain must 1) be a subdomain of the verified identity, 2) not be used in a "From" address if the MAIL FROM domain is the destination of email feedback forwarding (for more information, see the Amazon SES Developer Guide), and 3) not be used to receive emails. A value of null disables the custom MAIL FROM setting for the identity.

$sel:behaviorOnMXFailure:SetIdentityMailFromDomain', setIdentityMailFromDomain_behaviorOnMXFailure - The action that you want Amazon SES to take if it cannot successfully read the required MX record when you send an email. If you choose UseDefaultValue, Amazon SES will use amazonses.com (or a subdomain of that) as the MAIL FROM domain. If you choose RejectMessage, Amazon SES will return a MailFromDomainNotVerified error and not send the email.

The action specified in BehaviorOnMXFailure is taken when the custom MAIL FROM domain setup is in the Pending, Failed, and TemporaryFailure states.

$sel:identity:SetIdentityMailFromDomain', setIdentityMailFromDomain_identity - The verified identity for which you want to enable or disable the specified custom MAIL FROM domain.

Request Lenses

setIdentityMailFromDomain_mailFromDomain :: Lens' SetIdentityMailFromDomain (Maybe Text)

The custom MAIL FROM domain that you want the verified identity to use. The MAIL FROM domain must 1) be a subdomain of the verified identity, 2) not be used in a "From" address if the MAIL FROM domain is the destination of email feedback forwarding (for more information, see the Amazon SES Developer Guide), and 3) not be used to receive emails. A value of null disables the custom MAIL FROM setting for the identity.

setIdentityMailFromDomain_behaviorOnMXFailure :: Lens' SetIdentityMailFromDomain (Maybe BehaviorOnMXFailure)

The action that you want Amazon SES to take if it cannot successfully read the required MX record when you send an email. If you choose UseDefaultValue, Amazon SES will use amazonses.com (or a subdomain of that) as the MAIL FROM domain. If you choose RejectMessage, Amazon SES will return a MailFromDomainNotVerified error and not send the email.

The action specified in BehaviorOnMXFailure is taken when the custom MAIL FROM domain setup is in the Pending, Failed, and TemporaryFailure states.

setIdentityMailFromDomain_identity :: Lens' SetIdentityMailFromDomain Text

The verified identity for which you want to enable or disable the specified custom MAIL FROM domain.

Destructuring the Response

data SetIdentityMailFromDomainResponse

An empty element returned on a successful request.

See: newSetIdentityMailFromDomainResponse smart constructor.

Constructors

SetIdentityMailFromDomainResponse'

Fields

httpStatus :: Int

The response's http status code.

newSetIdentityMailFromDomainResponse

Arguments

:: Int
$sel:httpStatus:SetIdentityMailFromDomainResponse'

-> SetIdentityMailFromDomainResponse

Create a value of SetIdentityMailFromDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SetIdentityMailFromDomainResponse', setIdentityMailFromDomainResponse_httpStatus - The response's http status code.

Response Lenses

setIdentityMailFromDomainResponse_httpStatus :: Lens' SetIdentityMailFromDomainResponse Int

The response's http status code.