Amazonka.SES.SetReceiptRulePosition

Description

Sets the position of the specified receipt rule in the receipt rule set.

For information about managing receipt rules, see the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data SetReceiptRulePosition

Represents a request to set the position of a receipt rule in a receipt rule set. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

See: newSetReceiptRulePosition smart constructor.

Constructors

SetReceiptRulePosition'

Fields

after :: Maybe Text

The name of the receipt rule after which to place the specified receipt rule.

ruleSetName :: Text

The name of the receipt rule set that contains the receipt rule to reposition.

ruleName :: Text

The name of the receipt rule to reposition.

newSetReceiptRulePosition

Arguments

:: Text
$sel:ruleSetName:SetReceiptRulePosition'

-> Text
$sel:ruleName:SetReceiptRulePosition'

-> SetReceiptRulePosition

Create a value of SetReceiptRulePosition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:after:SetReceiptRulePosition', setReceiptRulePosition_after - The name of the receipt rule after which to place the specified receipt rule.

$sel:ruleSetName:SetReceiptRulePosition', setReceiptRulePosition_ruleSetName - The name of the receipt rule set that contains the receipt rule to reposition.

$sel:ruleName:SetReceiptRulePosition', setReceiptRulePosition_ruleName - The name of the receipt rule to reposition.

Request Lenses

setReceiptRulePosition_after :: Lens' SetReceiptRulePosition (Maybe Text)

The name of the receipt rule after which to place the specified receipt rule.

setReceiptRulePosition_ruleSetName :: Lens' SetReceiptRulePosition Text

The name of the receipt rule set that contains the receipt rule to reposition.

setReceiptRulePosition_ruleName :: Lens' SetReceiptRulePosition Text

The name of the receipt rule to reposition.

Destructuring the Response

data SetReceiptRulePositionResponse

An empty element returned on a successful request.

See: newSetReceiptRulePositionResponse smart constructor.

Constructors

SetReceiptRulePositionResponse'

Fields

httpStatus :: Int

The response's http status code.

newSetReceiptRulePositionResponse

Arguments

:: Int
$sel:httpStatus:SetReceiptRulePositionResponse'

-> SetReceiptRulePositionResponse

Create a value of SetReceiptRulePositionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SetReceiptRulePositionResponse', setReceiptRulePositionResponse_httpStatus - The response's http status code.

Response Lenses

setReceiptRulePositionResponse_httpStatus :: Lens' SetReceiptRulePositionResponse Int

The response's http status code.