Amazonka.SES.Types.AddHeaderAction

Description

Documentation

data AddHeaderAction

When included in a receipt rule, this action adds a header to the received email.

For information about adding a header using a receipt rule, see the Amazon SES Developer Guide.

See: newAddHeaderAction smart constructor.

Constructors

AddHeaderAction'

Fields

headerName :: Text

The name of the header to add. Must be between 1 and 50 characters, inclusive, and consist of alphanumeric (a-z, A-Z, 0-9) characters and dashes only.

headerValue :: Text

Must be less than 2048 characters, and must not contain newline characters ("\r" or "\n").

newAddHeaderAction

Arguments

:: Text
$sel:headerName:AddHeaderAction'

-> Text
$sel:headerValue:AddHeaderAction'

-> AddHeaderAction

Create a value of AddHeaderAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:headerName:AddHeaderAction', addHeaderAction_headerName - The name of the header to add. Must be between 1 and 50 characters, inclusive, and consist of alphanumeric (a-z, A-Z, 0-9) characters and dashes only.

$sel:headerValue:AddHeaderAction', addHeaderAction_headerValue - Must be less than 2048 characters, and must not contain newline characters ("\r" or "\n").

addHeaderAction_headerName :: Lens' AddHeaderAction Text

The name of the header to add. Must be between 1 and 50 characters, inclusive, and consist of alphanumeric (a-z, A-Z, 0-9) characters and dashes only.

addHeaderAction_headerValue :: Lens' AddHeaderAction Text

Must be less than 2048 characters, and must not contain newline characters ("\r" or "\n").