Amazonka.SES.Types.CustomVerificationEmailTemplate

Description

Documentation

data CustomVerificationEmailTemplate

Contains information about a custom verification email template.

See: newCustomVerificationEmailTemplate smart constructor.

Constructors

CustomVerificationEmailTemplate'

Fields

fromEmailAddress :: Maybe Text

The email address that the custom verification email is sent from.

templateName :: Maybe Text

The name of the custom verification email template.

failureRedirectionURL :: Maybe Text

The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

templateSubject :: Maybe Text

The subject line of the custom verification email.

successRedirectionURL :: Maybe Text

The URL that the recipient of the verification email is sent to if his or her address is successfully verified.

newCustomVerificationEmailTemplate :: CustomVerificationEmailTemplate

Create a value of CustomVerificationEmailTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromEmailAddress:CustomVerificationEmailTemplate', customVerificationEmailTemplate_fromEmailAddress - The email address that the custom verification email is sent from.

$sel:templateName:CustomVerificationEmailTemplate', customVerificationEmailTemplate_templateName - The name of the custom verification email template.

$sel:failureRedirectionURL:CustomVerificationEmailTemplate', customVerificationEmailTemplate_failureRedirectionURL - The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

$sel:templateSubject:CustomVerificationEmailTemplate', customVerificationEmailTemplate_templateSubject - The subject line of the custom verification email.

$sel:successRedirectionURL:CustomVerificationEmailTemplate', customVerificationEmailTemplate_successRedirectionURL - The URL that the recipient of the verification email is sent to if his or her address is successfully verified.

customVerificationEmailTemplate_fromEmailAddress :: Lens' CustomVerificationEmailTemplate (Maybe Text)

The email address that the custom verification email is sent from.

customVerificationEmailTemplate_templateName :: Lens' CustomVerificationEmailTemplate (Maybe Text)

The name of the custom verification email template.

customVerificationEmailTemplate_failureRedirectionURL :: Lens' CustomVerificationEmailTemplate (Maybe Text)

The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

customVerificationEmailTemplate_templateSubject :: Lens' CustomVerificationEmailTemplate (Maybe Text)

The subject line of the custom verification email.

customVerificationEmailTemplate_successRedirectionURL :: Lens' CustomVerificationEmailTemplate (Maybe Text)

The URL that the recipient of the verification email is sent to if his or her address is successfully verified.