Amazonka.SES.Types.DeliveryOptions

Description

Documentation

data DeliveryOptions

Specifies whether messages that use the configuration set are required to use Transport Layer Security (TLS).

See: newDeliveryOptions smart constructor.

Constructors

DeliveryOptions'

Fields

tlsPolicy :: Maybe TlsPolicy

Specifies whether messages that use the configuration set are required to use Transport Layer Security (TLS). If the value is Require, messages are only delivered if a TLS connection can be established. If the value is Optional, messages can be delivered in plain text if a TLS connection can't be established.

newDeliveryOptions :: DeliveryOptions

Create a value of DeliveryOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tlsPolicy:DeliveryOptions', deliveryOptions_tlsPolicy - Specifies whether messages that use the configuration set are required to use Transport Layer Security (TLS). If the value is Require, messages are only delivered if a TLS connection can be established. If the value is Optional, messages can be delivered in plain text if a TLS connection can't be established.

deliveryOptions_tlsPolicy :: Lens' DeliveryOptions (Maybe TlsPolicy)

Specifies whether messages that use the configuration set are required to use Transport Layer Security (TLS). If the value is Require, messages are only delivered if a TLS connection can be established. If the value is Optional, messages can be delivered in plain text if a TLS connection can't be established.