Amazonka.SES.Types.EventDestination

Description

Documentation

data EventDestination

Contains information about the event destination that the specified email sending events will be published to.

When you create or update an event destination, you must provide one, and only one, destination. The destination can be Amazon CloudWatch, Amazon Kinesis Firehose or Amazon Simple Notification Service (Amazon SNS).

Event destinations are associated with configuration sets, which enable you to publish email sending events to Amazon CloudWatch, Amazon Kinesis Firehose, or Amazon Simple Notification Service (Amazon SNS). For information about using configuration sets, see the Amazon SES Developer Guide.

See: newEventDestination smart constructor.

Constructors

EventDestination'

Fields

enabled :: Maybe Bool

Sets whether Amazon SES publishes events to this destination when you send an email with the associated configuration set. Set to true to enable publishing to this destination; set to false to prevent publishing to this destination. The default value is false.

kinesisFirehoseDestination :: Maybe KinesisFirehoseDestination

An object that contains the delivery stream ARN and the IAM role ARN associated with an Amazon Kinesis Firehose event destination.

cloudWatchDestination :: Maybe CloudWatchDestination

An object that contains the names, default values, and sources of the dimensions associated with an Amazon CloudWatch event destination.

sNSDestination :: Maybe SNSDestination

An object that contains the topic ARN associated with an Amazon Simple Notification Service (Amazon SNS) event destination.

name :: Text

The name of the event destination. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Contain less than 64 characters.

matchingEventTypes :: [EventType]

The type of email sending events to publish to the event destination.

newEventDestination

Arguments

:: Text
$sel:name:EventDestination'

-> EventDestination

Create a value of EventDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:EventDestination', eventDestination_enabled - Sets whether Amazon SES publishes events to this destination when you send an email with the associated configuration set. Set to true to enable publishing to this destination; set to false to prevent publishing to this destination. The default value is false.

$sel:kinesisFirehoseDestination:EventDestination', eventDestination_kinesisFirehoseDestination - An object that contains the delivery stream ARN and the IAM role ARN associated with an Amazon Kinesis Firehose event destination.

$sel:cloudWatchDestination:EventDestination', eventDestination_cloudWatchDestination - An object that contains the names, default values, and sources of the dimensions associated with an Amazon CloudWatch event destination.

$sel:sNSDestination:EventDestination', eventDestination_sNSDestination - An object that contains the topic ARN associated with an Amazon Simple Notification Service (Amazon SNS) event destination.

$sel:name:EventDestination', eventDestination_name - The name of the event destination. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Contain less than 64 characters.

$sel:matchingEventTypes:EventDestination', eventDestination_matchingEventTypes - The type of email sending events to publish to the event destination.

eventDestination_enabled :: Lens' EventDestination (Maybe Bool)

Sets whether Amazon SES publishes events to this destination when you send an email with the associated configuration set. Set to true to enable publishing to this destination; set to false to prevent publishing to this destination. The default value is false.

eventDestination_kinesisFirehoseDestination :: Lens' EventDestination (Maybe KinesisFirehoseDestination)

An object that contains the delivery stream ARN and the IAM role ARN associated with an Amazon Kinesis Firehose event destination.

eventDestination_cloudWatchDestination :: Lens' EventDestination (Maybe CloudWatchDestination)

An object that contains the names, default values, and sources of the dimensions associated with an Amazon CloudWatch event destination.

eventDestination_sNSDestination :: Lens' EventDestination (Maybe SNSDestination)

An object that contains the topic ARN associated with an Amazon Simple Notification Service (Amazon SNS) event destination.

eventDestination_name :: Lens' EventDestination Text

The name of the event destination. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Contain less than 64 characters.

eventDestination_matchingEventTypes :: Lens' EventDestination [EventType]

The type of email sending events to publish to the event destination.