Amazonka.SES.Types.IdentityMailFromDomainAttributes

Description

Documentation

data IdentityMailFromDomainAttributes

Represents the custom MAIL FROM domain attributes of a verified identity (email address or domain).

See: newIdentityMailFromDomainAttributes smart constructor.

Constructors

IdentityMailFromDomainAttributes'

Fields

mailFromDomain :: Text

The custom MAIL FROM domain that the identity is configured to use.

mailFromDomainStatus :: CustomMailFromStatus

The state that indicates whether Amazon SES has successfully read the MX record required for custom MAIL FROM domain setup. If the state is Success, Amazon SES uses the specified custom MAIL FROM domain when the verified identity sends an email. All other states indicate that Amazon SES takes the action described by BehaviorOnMXFailure.

behaviorOnMXFailure :: BehaviorOnMXFailure

The action that Amazon SES takes if it cannot successfully read the required MX record when you send an email. A value of UseDefaultValue indicates that if Amazon SES cannot read the required MX record, it uses amazonses.com (or a subdomain of that) as the MAIL FROM domain. A value of RejectMessage indicates that if Amazon SES cannot read the required MX record, Amazon SES returns a MailFromDomainNotVerified error and does not send the email.

The custom MAIL FROM setup states that result in this behavior are Pending, Failed, and TemporaryFailure.

newIdentityMailFromDomainAttributes

Arguments

:: Text
$sel:mailFromDomain:IdentityMailFromDomainAttributes'

-> CustomMailFromStatus
$sel:mailFromDomainStatus:IdentityMailFromDomainAttributes'

-> BehaviorOnMXFailure
$sel:behaviorOnMXFailure:IdentityMailFromDomainAttributes'

-> IdentityMailFromDomainAttributes

Create a value of IdentityMailFromDomainAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mailFromDomain:IdentityMailFromDomainAttributes', identityMailFromDomainAttributes_mailFromDomain - The custom MAIL FROM domain that the identity is configured to use.

$sel:mailFromDomainStatus:IdentityMailFromDomainAttributes', identityMailFromDomainAttributes_mailFromDomainStatus - The state that indicates whether Amazon SES has successfully read the MX record required for custom MAIL FROM domain setup. If the state is Success, Amazon SES uses the specified custom MAIL FROM domain when the verified identity sends an email. All other states indicate that Amazon SES takes the action described by BehaviorOnMXFailure.

$sel:behaviorOnMXFailure:IdentityMailFromDomainAttributes', identityMailFromDomainAttributes_behaviorOnMXFailure - The action that Amazon SES takes if it cannot successfully read the required MX record when you send an email. A value of UseDefaultValue indicates that if Amazon SES cannot read the required MX record, it uses amazonses.com (or a subdomain of that) as the MAIL FROM domain. A value of RejectMessage indicates that if Amazon SES cannot read the required MX record, Amazon SES returns a MailFromDomainNotVerified error and does not send the email.

The custom MAIL FROM setup states that result in this behavior are Pending, Failed, and TemporaryFailure.

identityMailFromDomainAttributes_mailFromDomain :: Lens' IdentityMailFromDomainAttributes Text

The custom MAIL FROM domain that the identity is configured to use.

identityMailFromDomainAttributes_mailFromDomainStatus :: Lens' IdentityMailFromDomainAttributes CustomMailFromStatus

The state that indicates whether Amazon SES has successfully read the MX record required for custom MAIL FROM domain setup. If the state is Success, Amazon SES uses the specified custom MAIL FROM domain when the verified identity sends an email. All other states indicate that Amazon SES takes the action described by BehaviorOnMXFailure.

identityMailFromDomainAttributes_behaviorOnMXFailure :: Lens' IdentityMailFromDomainAttributes BehaviorOnMXFailure

The action that Amazon SES takes if it cannot successfully read the required MX record when you send an email. A value of UseDefaultValue indicates that if Amazon SES cannot read the required MX record, it uses amazonses.com (or a subdomain of that) as the MAIL FROM domain. A value of RejectMessage indicates that if Amazon SES cannot read the required MX record, Amazon SES returns a MailFromDomainNotVerified error and does not send the email.

The custom MAIL FROM setup states that result in this behavior are Pending, Failed, and TemporaryFailure.