Amazonka.SES.Types.IdentityNotificationAttributes

Description

Documentation

data IdentityNotificationAttributes

Represents the notification attributes of an identity, including whether an identity has Amazon Simple Notification Service (Amazon SNS) topics set for bounce, complaint, and/or delivery notifications, and whether feedback forwarding is enabled for bounce and complaint notifications.

See: newIdentityNotificationAttributes smart constructor.

Constructors

IdentityNotificationAttributes'

Fields

headersInDeliveryNotificationsEnabled :: Maybe Bool

Describes whether Amazon SES includes the original email headers in Amazon SNS notifications of type Delivery. A value of true specifies that Amazon SES will include headers in delivery notifications, and a value of false specifies that Amazon SES will not include headers in delivery notifications.

headersInComplaintNotificationsEnabled :: Maybe Bool

Describes whether Amazon SES includes the original email headers in Amazon SNS notifications of type Complaint. A value of true specifies that Amazon SES will include headers in complaint notifications, and a value of false specifies that Amazon SES will not include headers in complaint notifications.

headersInBounceNotificationsEnabled :: Maybe Bool

Describes whether Amazon SES includes the original email headers in Amazon SNS notifications of type Bounce. A value of true specifies that Amazon SES will include headers in bounce notifications, and a value of false specifies that Amazon SES will not include headers in bounce notifications.

bounceTopic :: Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish bounce notifications.

complaintTopic :: Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish complaint notifications.

deliveryTopic :: Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish delivery notifications.

forwardingEnabled :: Bool

Describes whether Amazon SES will forward bounce and complaint notifications as email. true indicates that Amazon SES will forward bounce and complaint notifications as email, while false indicates that bounce and complaint notifications will be published only to the specified bounce and complaint Amazon SNS topics.

newIdentityNotificationAttributes

Arguments

:: Text
$sel:bounceTopic:IdentityNotificationAttributes'

-> Text
$sel:complaintTopic:IdentityNotificationAttributes'

-> Text
$sel:deliveryTopic:IdentityNotificationAttributes'

-> Bool
$sel:forwardingEnabled:IdentityNotificationAttributes'

-> IdentityNotificationAttributes

Create a value of IdentityNotificationAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:headersInDeliveryNotificationsEnabled:IdentityNotificationAttributes', identityNotificationAttributes_headersInDeliveryNotificationsEnabled - Describes whether Amazon SES includes the original email headers in Amazon SNS notifications of type Delivery. A value of true specifies that Amazon SES will include headers in delivery notifications, and a value of false specifies that Amazon SES will not include headers in delivery notifications.

$sel:headersInComplaintNotificationsEnabled:IdentityNotificationAttributes', identityNotificationAttributes_headersInComplaintNotificationsEnabled - Describes whether Amazon SES includes the original email headers in Amazon SNS notifications of type Complaint. A value of true specifies that Amazon SES will include headers in complaint notifications, and a value of false specifies that Amazon SES will not include headers in complaint notifications.

$sel:headersInBounceNotificationsEnabled:IdentityNotificationAttributes', identityNotificationAttributes_headersInBounceNotificationsEnabled - Describes whether Amazon SES includes the original email headers in Amazon SNS notifications of type Bounce. A value of true specifies that Amazon SES will include headers in bounce notifications, and a value of false specifies that Amazon SES will not include headers in bounce notifications.

$sel:bounceTopic:IdentityNotificationAttributes', identityNotificationAttributes_bounceTopic - The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish bounce notifications.

$sel:complaintTopic:IdentityNotificationAttributes', identityNotificationAttributes_complaintTopic - The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish complaint notifications.

$sel:deliveryTopic:IdentityNotificationAttributes', identityNotificationAttributes_deliveryTopic - The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish delivery notifications.

$sel:forwardingEnabled:IdentityNotificationAttributes', identityNotificationAttributes_forwardingEnabled - Describes whether Amazon SES will forward bounce and complaint notifications as email. true indicates that Amazon SES will forward bounce and complaint notifications as email, while false indicates that bounce and complaint notifications will be published only to the specified bounce and complaint Amazon SNS topics.

identityNotificationAttributes_headersInDeliveryNotificationsEnabled :: Lens' IdentityNotificationAttributes (Maybe Bool)

Describes whether Amazon SES includes the original email headers in Amazon SNS notifications of type Delivery. A value of true specifies that Amazon SES will include headers in delivery notifications, and a value of false specifies that Amazon SES will not include headers in delivery notifications.

identityNotificationAttributes_headersInComplaintNotificationsEnabled :: Lens' IdentityNotificationAttributes (Maybe Bool)

Describes whether Amazon SES includes the original email headers in Amazon SNS notifications of type Complaint. A value of true specifies that Amazon SES will include headers in complaint notifications, and a value of false specifies that Amazon SES will not include headers in complaint notifications.

identityNotificationAttributes_headersInBounceNotificationsEnabled :: Lens' IdentityNotificationAttributes (Maybe Bool)

Describes whether Amazon SES includes the original email headers in Amazon SNS notifications of type Bounce. A value of true specifies that Amazon SES will include headers in bounce notifications, and a value of false specifies that Amazon SES will not include headers in bounce notifications.

identityNotificationAttributes_bounceTopic :: Lens' IdentityNotificationAttributes Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish bounce notifications.

identityNotificationAttributes_complaintTopic :: Lens' IdentityNotificationAttributes Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish complaint notifications.

identityNotificationAttributes_deliveryTopic :: Lens' IdentityNotificationAttributes Text

The Amazon Resource Name (ARN) of the Amazon SNS topic where Amazon SES will publish delivery notifications.

identityNotificationAttributes_forwardingEnabled :: Lens' IdentityNotificationAttributes Bool

Describes whether Amazon SES will forward bounce and complaint notifications as email. true indicates that Amazon SES will forward bounce and complaint notifications as email, while false indicates that bounce and complaint notifications will be published only to the specified bounce and complaint Amazon SNS topics.