Amazonka.SES.Types.MessageTag

Description

Documentation

data MessageTag

Contains the name and value of a tag that you can provide to SendEmail or SendRawEmail to apply to an email.

Message tags, which you use with configuration sets, enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

See: newMessageTag smart constructor.

Constructors

MessageTag'

Fields

name :: Text

The name of the tag. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Contain less than 256 characters.

value :: Text

The value of the tag. The value must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Contain less than 256 characters.

newMessageTag

Arguments

:: Text
$sel:name:MessageTag'

-> Text
$sel:value:MessageTag'

-> MessageTag

Create a value of MessageTag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:MessageTag', messageTag_name - The name of the tag. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Contain less than 256 characters.

$sel:value:MessageTag', messageTag_value - The value of the tag. The value must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Contain less than 256 characters.

messageTag_name :: Lens' MessageTag Text

The name of the tag. The name must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Contain less than 256 characters.

messageTag_value :: Lens' MessageTag Text

The value of the tag. The value must:

This value can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

Contain less than 256 characters.