Amazonka.SES.Types.TrackingOptions

Description

Documentation

data TrackingOptions

A domain that is used to redirect email recipients to an Amazon SES-operated domain. This domain captures open and click events generated by Amazon SES emails.

For more information, see Configuring Custom Domains to Handle Open and Click Tracking in the Amazon SES Developer Guide.

See: newTrackingOptions smart constructor.

Constructors

TrackingOptions'

Fields

customRedirectDomain :: Maybe Text

The custom subdomain that will be used to redirect email recipients to the Amazon SES event tracking domain.

newTrackingOptions :: TrackingOptions

Create a value of TrackingOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customRedirectDomain:TrackingOptions', trackingOptions_customRedirectDomain - The custom subdomain that will be used to redirect email recipients to the Amazon SES event tracking domain.

trackingOptions_customRedirectDomain :: Lens' TrackingOptions (Maybe Text)

The custom subdomain that will be used to redirect email recipients to the Amazon SES event tracking domain.