Amazonka.SES.UpdateConfigurationSetSendingEnabled

Description

Enables or disables email sending for messages sent using a specific configuration set in a given AWS Region. You can use this operation in conjunction with Amazon CloudWatch alarms to temporarily pause email sending for a configuration set when the reputation metrics for that configuration set (such as your bounce on complaint rate) exceed certain thresholds.

You can execute this operation no more than once per second.

Creating a Request

data UpdateConfigurationSetSendingEnabled

Represents a request to enable or disable the email sending capabilities for a specific configuration set.

See: newUpdateConfigurationSetSendingEnabled smart constructor.

Constructors

UpdateConfigurationSetSendingEnabled'

Fields

configurationSetName :: Text

The name of the configuration set that you want to update.

enabled :: Bool

Describes whether email sending is enabled or disabled for the configuration set.

newUpdateConfigurationSetSendingEnabled

Arguments

:: Text
$sel:configurationSetName:UpdateConfigurationSetSendingEnabled'

-> Bool
$sel:enabled:UpdateConfigurationSetSendingEnabled'

-> UpdateConfigurationSetSendingEnabled

Create a value of UpdateConfigurationSetSendingEnabled with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetName:UpdateConfigurationSetSendingEnabled', updateConfigurationSetSendingEnabled_configurationSetName - The name of the configuration set that you want to update.

$sel:enabled:UpdateConfigurationSetSendingEnabled', updateConfigurationSetSendingEnabled_enabled - Describes whether email sending is enabled or disabled for the configuration set.

Request Lenses

updateConfigurationSetSendingEnabled_configurationSetName :: Lens' UpdateConfigurationSetSendingEnabled Text

The name of the configuration set that you want to update.

updateConfigurationSetSendingEnabled_enabled :: Lens' UpdateConfigurationSetSendingEnabled Bool

Describes whether email sending is enabled or disabled for the configuration set.

Destructuring the Response

data UpdateConfigurationSetSendingEnabledResponse

See: newUpdateConfigurationSetSendingEnabledResponse smart constructor.

Constructors

UpdateConfigurationSetSendingEnabledResponse'

newUpdateConfigurationSetSendingEnabledResponse :: UpdateConfigurationSetSendingEnabledResponse

Create a value of UpdateConfigurationSetSendingEnabledResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.