Amazonka.SES.VerifyEmailIdentity

Description

Adds an email address to the list of identities for your Amazon SES account in the current AWS region and attempts to verify it. As a result of executing this operation, a verification email is sent to the specified address.

You can execute this operation no more than once per second.

Creating a Request

data VerifyEmailIdentity

Represents a request to begin email address verification with Amazon SES. For information about email address verification, see the Amazon SES Developer Guide.

See: newVerifyEmailIdentity smart constructor.

Constructors

VerifyEmailIdentity'

Fields

emailAddress :: Text

The email address to be verified.

newVerifyEmailIdentity

Arguments

:: Text
$sel:emailAddress:VerifyEmailIdentity'

-> VerifyEmailIdentity

Create a value of VerifyEmailIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:emailAddress:VerifyEmailIdentity', verifyEmailIdentity_emailAddress - The email address to be verified.

Request Lenses

verifyEmailIdentity_emailAddress :: Lens' VerifyEmailIdentity Text

The email address to be verified.

Destructuring the Response

data VerifyEmailIdentityResponse

An empty element returned on a successful request.

See: newVerifyEmailIdentityResponse smart constructor.

Constructors

VerifyEmailIdentityResponse'

Fields

httpStatus :: Int

The response's http status code.

newVerifyEmailIdentityResponse

Arguments

:: Int
$sel:httpStatus:VerifyEmailIdentityResponse'

-> VerifyEmailIdentityResponse

Create a value of VerifyEmailIdentityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:VerifyEmailIdentityResponse', verifyEmailIdentityResponse_httpStatus - The response's http status code.

Response Lenses

verifyEmailIdentityResponse_httpStatus :: Lens' VerifyEmailIdentityResponse Int

The response's http status code.