Amazonka.SESV2.DeleteConfigurationSetEventDestination

Description

Delete an event destination.

Events include message sends, deliveries, opens, clicks, bounces, and complaints. Event destinations are places that you can send information about these events to. For example, you can send event data to Amazon SNS to receive notifications when you receive bounces or complaints, or you can use Amazon Kinesis Data Firehose to stream data to Amazon S3 for long-term storage.

Creating a Request

data DeleteConfigurationSetEventDestination

A request to delete an event destination from a configuration set.

See: newDeleteConfigurationSetEventDestination smart constructor.

Constructors

DeleteConfigurationSetEventDestination'

Fields

configurationSetName :: Text

The name of the configuration set that contains the event destination to delete.

eventDestinationName :: Text

The name of the event destination to delete.

newDeleteConfigurationSetEventDestination

Arguments

:: Text
$sel:configurationSetName:DeleteConfigurationSetEventDestination'

-> Text
$sel:eventDestinationName:DeleteConfigurationSetEventDestination'

-> DeleteConfigurationSetEventDestination

Create a value of DeleteConfigurationSetEventDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetName:DeleteConfigurationSetEventDestination', deleteConfigurationSetEventDestination_configurationSetName - The name of the configuration set that contains the event destination to delete.

$sel:eventDestinationName:DeleteConfigurationSetEventDestination', deleteConfigurationSetEventDestination_eventDestinationName - The name of the event destination to delete.

Request Lenses

deleteConfigurationSetEventDestination_configurationSetName :: Lens' DeleteConfigurationSetEventDestination Text

The name of the configuration set that contains the event destination to delete.

deleteConfigurationSetEventDestination_eventDestinationName :: Lens' DeleteConfigurationSetEventDestination Text

The name of the event destination to delete.

Destructuring the Response

data DeleteConfigurationSetEventDestinationResponse

An HTTP 200 response if the request succeeds, or an error message if the request fails.

See: newDeleteConfigurationSetEventDestinationResponse smart constructor.

Constructors

DeleteConfigurationSetEventDestinationResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteConfigurationSetEventDestinationResponse

Arguments

:: Int
$sel:httpStatus:DeleteConfigurationSetEventDestinationResponse'

-> DeleteConfigurationSetEventDestinationResponse

Create a value of DeleteConfigurationSetEventDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteConfigurationSetEventDestinationResponse', deleteConfigurationSetEventDestinationResponse_httpStatus - The response's http status code.

Response Lenses

deleteConfigurationSetEventDestinationResponse_httpStatus :: Lens' DeleteConfigurationSetEventDestinationResponse Int

The response's http status code.