Amazonka.SESV2.GetCustomVerificationEmailTemplate

Description

Returns the custom email verification template for the template name you specify.

For more information about custom verification email templates, see Using Custom Verification Email Templates in the Amazon SES Developer Guide.

You can execute this operation no more than once per second.

Creating a Request

data GetCustomVerificationEmailTemplate

Represents a request to retrieve an existing custom verification email template.

See: newGetCustomVerificationEmailTemplate smart constructor.

Constructors

GetCustomVerificationEmailTemplate'

Fields

templateName :: Text

The name of the custom verification email template that you want to retrieve.

newGetCustomVerificationEmailTemplate

Arguments

:: Text
$sel:templateName:GetCustomVerificationEmailTemplate'

-> GetCustomVerificationEmailTemplate

Create a value of GetCustomVerificationEmailTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:GetCustomVerificationEmailTemplate', getCustomVerificationEmailTemplate_templateName - The name of the custom verification email template that you want to retrieve.

Request Lenses

getCustomVerificationEmailTemplate_templateName :: Lens' GetCustomVerificationEmailTemplate Text

The name of the custom verification email template that you want to retrieve.

Destructuring the Response

data GetCustomVerificationEmailTemplateResponse

The following elements are returned by the service.

See: newGetCustomVerificationEmailTemplateResponse smart constructor.

Constructors

GetCustomVerificationEmailTemplateResponse'

Fields

fromEmailAddress :: Maybe Text

The email address that the custom verification email is sent from.

templateName :: Maybe Text

The name of the custom verification email template.

failureRedirectionURL :: Maybe Text

The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

templateSubject :: Maybe Text

The subject line of the custom verification email.

successRedirectionURL :: Maybe Text

The URL that the recipient of the verification email is sent to if his or her address is successfully verified.

templateContent :: Maybe Text

The content of the custom verification email.

httpStatus :: Int

The response's http status code.

newGetCustomVerificationEmailTemplateResponse

Arguments

:: Int
$sel:httpStatus:GetCustomVerificationEmailTemplateResponse'

-> GetCustomVerificationEmailTemplateResponse

Create a value of GetCustomVerificationEmailTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromEmailAddress:GetCustomVerificationEmailTemplateResponse', getCustomVerificationEmailTemplateResponse_fromEmailAddress - The email address that the custom verification email is sent from.

$sel:templateName:GetCustomVerificationEmailTemplate', getCustomVerificationEmailTemplateResponse_templateName - The name of the custom verification email template.

$sel:failureRedirectionURL:GetCustomVerificationEmailTemplateResponse', getCustomVerificationEmailTemplateResponse_failureRedirectionURL - The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

$sel:templateSubject:GetCustomVerificationEmailTemplateResponse', getCustomVerificationEmailTemplateResponse_templateSubject - The subject line of the custom verification email.

$sel:successRedirectionURL:GetCustomVerificationEmailTemplateResponse', getCustomVerificationEmailTemplateResponse_successRedirectionURL - The URL that the recipient of the verification email is sent to if his or her address is successfully verified.

$sel:templateContent:GetCustomVerificationEmailTemplateResponse', getCustomVerificationEmailTemplateResponse_templateContent - The content of the custom verification email.

$sel:httpStatus:GetCustomVerificationEmailTemplateResponse', getCustomVerificationEmailTemplateResponse_httpStatus - The response's http status code.

Response Lenses

getCustomVerificationEmailTemplateResponse_fromEmailAddress :: Lens' GetCustomVerificationEmailTemplateResponse (Maybe Text)

The email address that the custom verification email is sent from.

getCustomVerificationEmailTemplateResponse_templateName :: Lens' GetCustomVerificationEmailTemplateResponse (Maybe Text)

The name of the custom verification email template.

getCustomVerificationEmailTemplateResponse_failureRedirectionURL :: Lens' GetCustomVerificationEmailTemplateResponse (Maybe Text)

The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

getCustomVerificationEmailTemplateResponse_templateSubject :: Lens' GetCustomVerificationEmailTemplateResponse (Maybe Text)

The subject line of the custom verification email.

getCustomVerificationEmailTemplateResponse_successRedirectionURL :: Lens' GetCustomVerificationEmailTemplateResponse (Maybe Text)

The URL that the recipient of the verification email is sent to if his or her address is successfully verified.

getCustomVerificationEmailTemplateResponse_templateContent :: Lens' GetCustomVerificationEmailTemplateResponse (Maybe Text)

The content of the custom verification email.

getCustomVerificationEmailTemplateResponse_httpStatus :: Lens' GetCustomVerificationEmailTemplateResponse Int

The response's http status code.