Amazonka.SESV2.ListDeliverabilityTestReports

Description

Show a list of the predictive inbox placement tests that you've performed, regardless of their statuses. For predictive inbox placement tests that are complete, you can use the GetDeliverabilityTestReport operation to view the results.

Creating a Request

data ListDeliverabilityTestReports

A request to list all of the predictive inbox placement tests that you've performed.

See: newListDeliverabilityTestReports smart constructor.

Constructors

ListDeliverabilityTestReports'

Fields

nextToken :: Maybe Text

A token returned from a previous call to ListDeliverabilityTestReports to indicate the position in the list of predictive inbox placement tests.

pageSize :: Maybe Int

The number of results to show in a single call to ListDeliverabilityTestReports. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

The value you specify has to be at least 0, and can be no more than 1000.

newListDeliverabilityTestReports :: ListDeliverabilityTestReports

Create a value of ListDeliverabilityTestReports with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDeliverabilityTestReports', listDeliverabilityTestReports_nextToken - A token returned from a previous call to ListDeliverabilityTestReports to indicate the position in the list of predictive inbox placement tests.

$sel:pageSize:ListDeliverabilityTestReports', listDeliverabilityTestReports_pageSize - The number of results to show in a single call to ListDeliverabilityTestReports. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

The value you specify has to be at least 0, and can be no more than 1000.

Request Lenses

listDeliverabilityTestReports_nextToken :: Lens' ListDeliverabilityTestReports (Maybe Text)

A token returned from a previous call to ListDeliverabilityTestReports to indicate the position in the list of predictive inbox placement tests.

listDeliverabilityTestReports_pageSize :: Lens' ListDeliverabilityTestReports (Maybe Int)

The number of results to show in a single call to ListDeliverabilityTestReports. If the number of results is larger than the number you specified in this parameter, then the response includes a NextToken element, which you can use to obtain additional results.

The value you specify has to be at least 0, and can be no more than 1000.

Destructuring the Response

data ListDeliverabilityTestReportsResponse

A list of the predictive inbox placement test reports that are available for your account, regardless of whether or not those tests are complete.

See: newListDeliverabilityTestReportsResponse smart constructor.

Constructors

ListDeliverabilityTestReportsResponse'

Fields

nextToken :: Maybe Text

A token that indicates that there are additional predictive inbox placement tests to list. To view additional predictive inbox placement tests, issue another request to ListDeliverabilityTestReports, and pass this token in the NextToken parameter.

httpStatus :: Int

The response's http status code.

deliverabilityTestReports :: [DeliverabilityTestReport]

An object that contains a lists of predictive inbox placement tests that you've performed.

newListDeliverabilityTestReportsResponse

Arguments

:: Int
$sel:httpStatus:ListDeliverabilityTestReportsResponse'

-> ListDeliverabilityTestReportsResponse

Create a value of ListDeliverabilityTestReportsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDeliverabilityTestReports', listDeliverabilityTestReportsResponse_nextToken - A token that indicates that there are additional predictive inbox placement tests to list. To view additional predictive inbox placement tests, issue another request to ListDeliverabilityTestReports, and pass this token in the NextToken parameter.

$sel:httpStatus:ListDeliverabilityTestReportsResponse', listDeliverabilityTestReportsResponse_httpStatus - The response's http status code.

$sel:deliverabilityTestReports:ListDeliverabilityTestReportsResponse', listDeliverabilityTestReportsResponse_deliverabilityTestReports - An object that contains a lists of predictive inbox placement tests that you've performed.

Response Lenses

listDeliverabilityTestReportsResponse_nextToken :: Lens' ListDeliverabilityTestReportsResponse (Maybe Text)

A token that indicates that there are additional predictive inbox placement tests to list. To view additional predictive inbox placement tests, issue another request to ListDeliverabilityTestReports, and pass this token in the NextToken parameter.

listDeliverabilityTestReportsResponse_httpStatus :: Lens' ListDeliverabilityTestReportsResponse Int

The response's http status code.

listDeliverabilityTestReportsResponse_deliverabilityTestReports :: Lens' ListDeliverabilityTestReportsResponse [DeliverabilityTestReport]

An object that contains a lists of predictive inbox placement tests that you've performed.