Amazonka.SESV2.PutEmailIdentityDkimSigningAttributes

Description

Used to configure or change the DKIM authentication settings for an email domain identity. You can use this operation to do any of the following:

Update the signing attributes for an identity that uses Bring Your Own DKIM (BYODKIM).

Update the key length that should be used for Easy DKIM.

Change from using no DKIM authentication to using Easy DKIM.

Change from using no DKIM authentication to using BYODKIM.

Change from using Easy DKIM to using BYODKIM.

Change from using BYODKIM to using Easy DKIM.

Creating a Request

data PutEmailIdentityDkimSigningAttributes

A request to change the DKIM attributes for an email identity.

See: newPutEmailIdentityDkimSigningAttributes smart constructor.

Constructors

PutEmailIdentityDkimSigningAttributes'

Fields

signingAttributes :: Maybe DkimSigningAttributes

An object that contains information about the private key and selector that you want to use to configure DKIM for the identity for Bring Your Own DKIM (BYODKIM) for the identity, or, configures the key length to be used for Easy DKIM.

emailIdentity :: Text

The email identity.

signingAttributesOrigin :: DkimSigningAttributesOrigin

The method to use to configure DKIM for the identity. There are the following possible values:

AWS_SES – Configure DKIM for the identity by using Easy DKIM.

EXTERNAL – Configure DKIM for the identity by using Bring Your Own DKIM (BYODKIM).

newPutEmailIdentityDkimSigningAttributes

Arguments

:: Text
$sel:emailIdentity:PutEmailIdentityDkimSigningAttributes'

-> DkimSigningAttributesOrigin
$sel:signingAttributesOrigin:PutEmailIdentityDkimSigningAttributes'

-> PutEmailIdentityDkimSigningAttributes

Create a value of PutEmailIdentityDkimSigningAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signingAttributes:PutEmailIdentityDkimSigningAttributes', putEmailIdentityDkimSigningAttributes_signingAttributes - An object that contains information about the private key and selector that you want to use to configure DKIM for the identity for Bring Your Own DKIM (BYODKIM) for the identity, or, configures the key length to be used for Easy DKIM.

$sel:emailIdentity:PutEmailIdentityDkimSigningAttributes', putEmailIdentityDkimSigningAttributes_emailIdentity - The email identity.

$sel:signingAttributesOrigin:PutEmailIdentityDkimSigningAttributes', putEmailIdentityDkimSigningAttributes_signingAttributesOrigin - The method to use to configure DKIM for the identity. There are the following possible values:

AWS_SES – Configure DKIM for the identity by using Easy DKIM.

EXTERNAL – Configure DKIM for the identity by using Bring Your Own DKIM (BYODKIM).

Request Lenses

putEmailIdentityDkimSigningAttributes_signingAttributes :: Lens' PutEmailIdentityDkimSigningAttributes (Maybe DkimSigningAttributes)

An object that contains information about the private key and selector that you want to use to configure DKIM for the identity for Bring Your Own DKIM (BYODKIM) for the identity, or, configures the key length to be used for Easy DKIM.

putEmailIdentityDkimSigningAttributes_emailIdentity :: Lens' PutEmailIdentityDkimSigningAttributes Text

The email identity.

putEmailIdentityDkimSigningAttributes_signingAttributesOrigin :: Lens' PutEmailIdentityDkimSigningAttributes DkimSigningAttributesOrigin

The method to use to configure DKIM for the identity. There are the following possible values:

AWS_SES – Configure DKIM for the identity by using Easy DKIM.

EXTERNAL – Configure DKIM for the identity by using Bring Your Own DKIM (BYODKIM).

Destructuring the Response

data PutEmailIdentityDkimSigningAttributesResponse

If the action is successful, the service sends back an HTTP 200 response.

The following data is returned in JSON format by the service.

See: newPutEmailIdentityDkimSigningAttributesResponse smart constructor.

Constructors

PutEmailIdentityDkimSigningAttributesResponse'

Fields

dkimStatus :: Maybe DkimStatus

The DKIM authentication status of the identity. Amazon SES determines the authentication status by searching for specific records in the DNS configuration for your domain. If you used Easy DKIM to set up DKIM authentication, Amazon SES tries to find three unique CNAME records in the DNS configuration for your domain.

If you provided a public key to perform DKIM authentication, Amazon SES tries to find a TXT record that uses the selector that you specified. The value of the TXT record must be a public key that's paired with the private key that you specified in the process of creating the identity.

The status can be one of the following:

PENDING – The verification process was initiated, but Amazon SES hasn't yet detected the DKIM records in the DNS configuration for the domain.

SUCCESS – The verification process completed successfully.

FAILED – The verification process failed. This typically occurs when Amazon SES fails to find the DKIM records in the DNS configuration of the domain.

TEMPORARY_FAILURE – A temporary issue is preventing Amazon SES from determining the DKIM authentication status of the domain.

NOT_STARTED – The DKIM verification process hasn't been initiated for the domain.

dkimTokens :: Maybe [Text]

If you used Easy DKIM to configure DKIM authentication for the domain, then this object contains a set of unique strings that you use to create a set of CNAME records that you add to the DNS configuration for your domain. When Amazon SES detects these records in the DNS configuration for your domain, the DKIM authentication process is complete.

If you configured DKIM authentication for the domain by providing your own public-private key pair, then this object contains the selector that's associated with your public key.

Regardless of the DKIM authentication method you use, Amazon SES searches for the appropriate records in the DNS configuration of the domain for up to 72 hours.

httpStatus :: Int

The response's http status code.

newPutEmailIdentityDkimSigningAttributesResponse

Arguments

:: Int
$sel:httpStatus:PutEmailIdentityDkimSigningAttributesResponse'

-> PutEmailIdentityDkimSigningAttributesResponse

Create a value of PutEmailIdentityDkimSigningAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dkimStatus:PutEmailIdentityDkimSigningAttributesResponse', putEmailIdentityDkimSigningAttributesResponse_dkimStatus - The DKIM authentication status of the identity. Amazon SES determines the authentication status by searching for specific records in the DNS configuration for your domain. If you used Easy DKIM to set up DKIM authentication, Amazon SES tries to find three unique CNAME records in the DNS configuration for your domain.

If you provided a public key to perform DKIM authentication, Amazon SES tries to find a TXT record that uses the selector that you specified. The value of the TXT record must be a public key that's paired with the private key that you specified in the process of creating the identity.

The status can be one of the following:

PENDING – The verification process was initiated, but Amazon SES hasn't yet detected the DKIM records in the DNS configuration for the domain.

SUCCESS – The verification process completed successfully.

FAILED – The verification process failed. This typically occurs when Amazon SES fails to find the DKIM records in the DNS configuration of the domain.

TEMPORARY_FAILURE – A temporary issue is preventing Amazon SES from determining the DKIM authentication status of the domain.

NOT_STARTED – The DKIM verification process hasn't been initiated for the domain.

$sel:dkimTokens:PutEmailIdentityDkimSigningAttributesResponse', putEmailIdentityDkimSigningAttributesResponse_dkimTokens - If you used Easy DKIM to configure DKIM authentication for the domain, then this object contains a set of unique strings that you use to create a set of CNAME records that you add to the DNS configuration for your domain. When Amazon SES detects these records in the DNS configuration for your domain, the DKIM authentication process is complete.

If you configured DKIM authentication for the domain by providing your own public-private key pair, then this object contains the selector that's associated with your public key.

Regardless of the DKIM authentication method you use, Amazon SES searches for the appropriate records in the DNS configuration of the domain for up to 72 hours.

$sel:httpStatus:PutEmailIdentityDkimSigningAttributesResponse', putEmailIdentityDkimSigningAttributesResponse_httpStatus - The response's http status code.

Response Lenses

putEmailIdentityDkimSigningAttributesResponse_dkimStatus :: Lens' PutEmailIdentityDkimSigningAttributesResponse (Maybe DkimStatus)

The DKIM authentication status of the identity. Amazon SES determines the authentication status by searching for specific records in the DNS configuration for your domain. If you used Easy DKIM to set up DKIM authentication, Amazon SES tries to find three unique CNAME records in the DNS configuration for your domain.

If you provided a public key to perform DKIM authentication, Amazon SES tries to find a TXT record that uses the selector that you specified. The value of the TXT record must be a public key that's paired with the private key that you specified in the process of creating the identity.

The status can be one of the following:

PENDING – The verification process was initiated, but Amazon SES hasn't yet detected the DKIM records in the DNS configuration for the domain.

SUCCESS – The verification process completed successfully.

FAILED – The verification process failed. This typically occurs when Amazon SES fails to find the DKIM records in the DNS configuration of the domain.

TEMPORARY_FAILURE – A temporary issue is preventing Amazon SES from determining the DKIM authentication status of the domain.

NOT_STARTED – The DKIM verification process hasn't been initiated for the domain.

putEmailIdentityDkimSigningAttributesResponse_dkimTokens :: Lens' PutEmailIdentityDkimSigningAttributesResponse (Maybe [Text])

If you used Easy DKIM to configure DKIM authentication for the domain, then this object contains a set of unique strings that you use to create a set of CNAME records that you add to the DNS configuration for your domain. When Amazon SES detects these records in the DNS configuration for your domain, the DKIM authentication process is complete.

If you configured DKIM authentication for the domain by providing your own public-private key pair, then this object contains the selector that's associated with your public key.

Regardless of the DKIM authentication method you use, Amazon SES searches for the appropriate records in the DNS configuration of the domain for up to 72 hours.

putEmailIdentityDkimSigningAttributesResponse_httpStatus :: Lens' PutEmailIdentityDkimSigningAttributesResponse Int

The response's http status code.