Amazonka.SESV2.SendBulkEmail

Description

Composes an email message to multiple destinations.

Creating a Request

data SendBulkEmail

Represents a request to send email messages to multiple destinations using Amazon SES. For more information, see the Amazon SES Developer Guide.

See: newSendBulkEmail smart constructor.

Constructors

SendBulkEmail'

Fields

fromEmailAddress :: Maybe Text

The email address to use as the "From" address for the email. The address that you specify has to be verified.

fromEmailAddressIdentityArn :: Maybe Text

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the FromEmailAddress parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use sender@example.com, then you would specify the FromEmailAddressIdentityArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the FromEmailAddress to be sender@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

configurationSetName :: Maybe Text

The name of the configuration set to use when sending the email.

feedbackForwardingEmailAddress :: Maybe Text

The address that you want bounce and complaint notifications to be sent to.

feedbackForwardingEmailAddressIdentityArn :: Maybe Text

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the FeedbackForwardingEmailAddress parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use feedback@example.com, then you would specify the FeedbackForwardingEmailAddressIdentityArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the FeedbackForwardingEmailAddress to be feedback@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

defaultEmailTags :: Maybe [MessageTag]

A list of tags, in the form of name/value pairs, to apply to an email that you send using the SendEmail operation. Tags correspond to characteristics of the email that you define, so that you can publish email sending events.

replyToAddresses :: Maybe [Text]

The "Reply-to" email addresses for the message. When the recipient replies to the message, each Reply-to address receives the reply.

defaultContent :: BulkEmailContent

An object that contains the body of the message. You can specify a template message.

bulkEmailEntries :: [BulkEmailEntry]

The list of bulk email entry objects.

newSendBulkEmail

Arguments

:: BulkEmailContent
$sel:defaultContent:SendBulkEmail'

-> SendBulkEmail

Create a value of SendBulkEmail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromEmailAddress:SendBulkEmail', sendBulkEmail_fromEmailAddress - The email address to use as the "From" address for the email. The address that you specify has to be verified.

$sel:fromEmailAddressIdentityArn:SendBulkEmail', sendBulkEmail_fromEmailAddressIdentityArn - This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the FromEmailAddress parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use sender@example.com, then you would specify the FromEmailAddressIdentityArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the FromEmailAddress to be sender@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

$sel:configurationSetName:SendBulkEmail', sendBulkEmail_configurationSetName - The name of the configuration set to use when sending the email.

$sel:feedbackForwardingEmailAddress:SendBulkEmail', sendBulkEmail_feedbackForwardingEmailAddress - The address that you want bounce and complaint notifications to be sent to.

$sel:feedbackForwardingEmailAddressIdentityArn:SendBulkEmail', sendBulkEmail_feedbackForwardingEmailAddressIdentityArn - This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the FeedbackForwardingEmailAddress parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use feedback@example.com, then you would specify the FeedbackForwardingEmailAddressIdentityArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the FeedbackForwardingEmailAddress to be feedback@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

$sel:defaultEmailTags:SendBulkEmail', sendBulkEmail_defaultEmailTags - A list of tags, in the form of name/value pairs, to apply to an email that you send using the SendEmail operation. Tags correspond to characteristics of the email that you define, so that you can publish email sending events.

$sel:replyToAddresses:SendBulkEmail', sendBulkEmail_replyToAddresses - The "Reply-to" email addresses for the message. When the recipient replies to the message, each Reply-to address receives the reply.

$sel:defaultContent:SendBulkEmail', sendBulkEmail_defaultContent - An object that contains the body of the message. You can specify a template message.

$sel:bulkEmailEntries:SendBulkEmail', sendBulkEmail_bulkEmailEntries - The list of bulk email entry objects.

Request Lenses

sendBulkEmail_fromEmailAddress :: Lens' SendBulkEmail (Maybe Text)

The email address to use as the "From" address for the email. The address that you specify has to be verified.

sendBulkEmail_fromEmailAddressIdentityArn :: Lens' SendBulkEmail (Maybe Text)

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the FromEmailAddress parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use sender@example.com, then you would specify the FromEmailAddressIdentityArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the FromEmailAddress to be sender@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

sendBulkEmail_configurationSetName :: Lens' SendBulkEmail (Maybe Text)

The name of the configuration set to use when sending the email.

sendBulkEmail_feedbackForwardingEmailAddress :: Lens' SendBulkEmail (Maybe Text)

The address that you want bounce and complaint notifications to be sent to.

sendBulkEmail_feedbackForwardingEmailAddressIdentityArn :: Lens' SendBulkEmail (Maybe Text)

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the FeedbackForwardingEmailAddress parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use feedback@example.com, then you would specify the FeedbackForwardingEmailAddressIdentityArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the FeedbackForwardingEmailAddress to be feedback@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

sendBulkEmail_defaultEmailTags :: Lens' SendBulkEmail (Maybe [MessageTag])

A list of tags, in the form of name/value pairs, to apply to an email that you send using the SendEmail operation. Tags correspond to characteristics of the email that you define, so that you can publish email sending events.

sendBulkEmail_replyToAddresses :: Lens' SendBulkEmail (Maybe [Text])

The "Reply-to" email addresses for the message. When the recipient replies to the message, each Reply-to address receives the reply.

sendBulkEmail_defaultContent :: Lens' SendBulkEmail BulkEmailContent

An object that contains the body of the message. You can specify a template message.

sendBulkEmail_bulkEmailEntries :: Lens' SendBulkEmail [BulkEmailEntry]

The list of bulk email entry objects.

Destructuring the Response

data SendBulkEmailResponse

The following data is returned in JSON format by the service.

See: newSendBulkEmailResponse smart constructor.

Constructors

SendBulkEmailResponse'

Fields

httpStatus :: Int

The response's http status code.

bulkEmailEntryResults :: [BulkEmailEntryResult]

One object per intended recipient. Check each response object and retry any messages with a failure status.

newSendBulkEmailResponse

Arguments

:: Int
$sel:httpStatus:SendBulkEmailResponse'

-> SendBulkEmailResponse

Create a value of SendBulkEmailResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SendBulkEmailResponse', sendBulkEmailResponse_httpStatus - The response's http status code.

$sel:bulkEmailEntryResults:SendBulkEmailResponse', sendBulkEmailResponse_bulkEmailEntryResults - One object per intended recipient. Check each response object and retry any messages with a failure status.

Response Lenses

sendBulkEmailResponse_httpStatus :: Lens' SendBulkEmailResponse Int

The response's http status code.

sendBulkEmailResponse_bulkEmailEntryResults :: Lens' SendBulkEmailResponse [BulkEmailEntryResult]

One object per intended recipient. Check each response object and retry any messages with a failure status.