Amazonka.SESV2.Types.AccountDetails

Description

Documentation

data AccountDetails

An object that contains information about your account details.

See: newAccountDetails smart constructor.

Constructors

AccountDetails'

Fields

reviewDetails :: Maybe ReviewDetails

Information about the review of the latest details you submitted.

mailType :: Maybe MailType

The type of email your account is sending. The mail type can be one of the following:

MARKETING – Most of your sending traffic is to keep your customers informed of your latest offering.

TRANSACTIONAL – Most of your sending traffic is to communicate during a transaction with a customer.

useCaseDescription :: Maybe (Sensitive Text)

A description of the types of email that you plan to send.

contactLanguage :: Maybe ContactLanguage

The language you would prefer for the case. The contact language can be one of ENGLISH or JAPANESE.

additionalContactEmailAddresses :: Maybe (Sensitive (NonEmpty (Sensitive Text)))

Additional email addresses where updates are sent about your account review process.

websiteURL :: Maybe (Sensitive Text)

The URL of your website. This information helps us better understand the type of content that you plan to send.

newAccountDetails :: AccountDetails

Create a value of AccountDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reviewDetails:AccountDetails', accountDetails_reviewDetails - Information about the review of the latest details you submitted.

$sel:mailType:AccountDetails', accountDetails_mailType - The type of email your account is sending. The mail type can be one of the following:

MARKETING – Most of your sending traffic is to keep your customers informed of your latest offering.

TRANSACTIONAL – Most of your sending traffic is to communicate during a transaction with a customer.

$sel:useCaseDescription:AccountDetails', accountDetails_useCaseDescription - A description of the types of email that you plan to send.

$sel:contactLanguage:AccountDetails', accountDetails_contactLanguage - The language you would prefer for the case. The contact language can be one of ENGLISH or JAPANESE.

$sel:additionalContactEmailAddresses:AccountDetails', accountDetails_additionalContactEmailAddresses - Additional email addresses where updates are sent about your account review process.

$sel:websiteURL:AccountDetails', accountDetails_websiteURL - The URL of your website. This information helps us better understand the type of content that you plan to send.

accountDetails_reviewDetails :: Lens' AccountDetails (Maybe ReviewDetails)

Information about the review of the latest details you submitted.

accountDetails_mailType :: Lens' AccountDetails (Maybe MailType)

The type of email your account is sending. The mail type can be one of the following:

MARKETING – Most of your sending traffic is to keep your customers informed of your latest offering.

TRANSACTIONAL – Most of your sending traffic is to communicate during a transaction with a customer.

accountDetails_useCaseDescription :: Lens' AccountDetails (Maybe Text)

A description of the types of email that you plan to send.

accountDetails_contactLanguage :: Lens' AccountDetails (Maybe ContactLanguage)

The language you would prefer for the case. The contact language can be one of ENGLISH or JAPANESE.

accountDetails_additionalContactEmailAddresses :: Lens' AccountDetails (Maybe (NonEmpty Text))

Additional email addresses where updates are sent about your account review process.

accountDetails_websiteURL :: Lens' AccountDetails (Maybe Text)

The URL of your website. This information helps us better understand the type of content that you plan to send.