Amazonka.SESV2.Types.ImportJobSummary

Description

Documentation

data ImportJobSummary

A summary of the import job.

See: newImportJobSummary smart constructor.

Constructors

ImportJobSummary'

Fields

jobId :: Maybe Text

importDestination :: Maybe ImportDestination

jobStatus :: Maybe JobStatus

createdTimestamp :: Maybe POSIX

The date and time when the import job was created.

newImportJobSummary :: ImportJobSummary

Create a value of ImportJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:ImportJobSummary', importJobSummary_jobId - Undocumented member.

$sel:importDestination:ImportJobSummary', importJobSummary_importDestination - Undocumented member.

$sel:jobStatus:ImportJobSummary', importJobSummary_jobStatus - Undocumented member.

$sel:createdTimestamp:ImportJobSummary', importJobSummary_createdTimestamp - The date and time when the import job was created.

importJobSummary_jobId :: Lens' ImportJobSummary (Maybe Text)

Undocumented member.

importJobSummary_importDestination :: Lens' ImportJobSummary (Maybe ImportDestination)

Undocumented member.

importJobSummary_jobStatus :: Lens' ImportJobSummary (Maybe JobStatus)

Undocumented member.

importJobSummary_createdTimestamp :: Lens' ImportJobSummary (Maybe UTCTime)

The date and time when the import job was created.