Amazonka.SESV2.Types.Topic

Description

Documentation

data Topic

An interest group, theme, or label within a list. Lists can have multiple topics.

See: newTopic smart constructor.

Constructors

Topic'

Fields

description :: Maybe Text

A description of what the topic is about, which the contact will see.

topicName :: Text

The name of the topic.

displayName :: Text

The name of the topic the contact will see.

defaultSubscriptionStatus :: SubscriptionStatus

The default subscription status to be applied to a contact if the contact has not noted their preference for subscribing to a topic.

newTopic

Arguments

:: Text
$sel:topicName:Topic'

-> Text
$sel:displayName:Topic'

-> SubscriptionStatus
$sel:defaultSubscriptionStatus:Topic'

-> Topic

Create a value of Topic with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:Topic', topic_description - A description of what the topic is about, which the contact will see.

$sel:topicName:Topic', topic_topicName - The name of the topic.

$sel:displayName:Topic', topic_displayName - The name of the topic the contact will see.

$sel:defaultSubscriptionStatus:Topic', topic_defaultSubscriptionStatus - The default subscription status to be applied to a contact if the contact has not noted their preference for subscribing to a topic.

topic_description :: Lens' Topic (Maybe Text)

A description of what the topic is about, which the contact will see.

topic_topicName :: Lens' Topic Text

The name of the topic.

topic_displayName :: Lens' Topic Text

The name of the topic the contact will see.

topic_defaultSubscriptionStatus :: Lens' Topic SubscriptionStatus

The default subscription status to be applied to a contact if the contact has not noted their preference for subscribing to a topic.