Amazonka.SMS.DeleteAppLaunchConfiguration

Description

Deletes the launch configuration for the specified application.

Creating a Request

data DeleteAppLaunchConfiguration

See: newDeleteAppLaunchConfiguration smart constructor.

Constructors

DeleteAppLaunchConfiguration'

Fields

appId :: Maybe Text

The ID of the application.

newDeleteAppLaunchConfiguration :: DeleteAppLaunchConfiguration

Create a value of DeleteAppLaunchConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:DeleteAppLaunchConfiguration', deleteAppLaunchConfiguration_appId - The ID of the application.

Request Lenses

deleteAppLaunchConfiguration_appId :: Lens' DeleteAppLaunchConfiguration (Maybe Text)

The ID of the application.

Destructuring the Response

data DeleteAppLaunchConfigurationResponse

See: newDeleteAppLaunchConfigurationResponse smart constructor.

Constructors

DeleteAppLaunchConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteAppLaunchConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DeleteAppLaunchConfigurationResponse'

-> DeleteAppLaunchConfigurationResponse

Create a value of DeleteAppLaunchConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteAppLaunchConfigurationResponse', deleteAppLaunchConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

deleteAppLaunchConfigurationResponse_httpStatus :: Lens' DeleteAppLaunchConfigurationResponse Int

The response's http status code.