Amazonka.SMS.GetAppLaunchConfiguration

Description

Retrieves the application launch configuration associated with the specified application.

Creating a Request

data GetAppLaunchConfiguration

See: newGetAppLaunchConfiguration smart constructor.

Constructors

GetAppLaunchConfiguration'

Fields

appId :: Maybe Text

The ID of the application.

newGetAppLaunchConfiguration :: GetAppLaunchConfiguration

Create a value of GetAppLaunchConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:GetAppLaunchConfiguration', getAppLaunchConfiguration_appId - The ID of the application.

Request Lenses

getAppLaunchConfiguration_appId :: Lens' GetAppLaunchConfiguration (Maybe Text)

The ID of the application.

Destructuring the Response

data GetAppLaunchConfigurationResponse

See: newGetAppLaunchConfigurationResponse smart constructor.

Constructors

GetAppLaunchConfigurationResponse'

Fields

serverGroupLaunchConfigurations :: Maybe [ServerGroupLaunchConfiguration]

The launch configurations for server groups in this application.

autoLaunch :: Maybe Bool

Indicates whether the application is configured to launch automatically after replication is complete.

roleName :: Maybe Text

The name of the service role in the customer's account that AWS CloudFormation uses to launch the application.

appId :: Maybe Text

The ID of the application.

httpStatus :: Int

The response's http status code.

newGetAppLaunchConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetAppLaunchConfigurationResponse'

-> GetAppLaunchConfigurationResponse

Create a value of GetAppLaunchConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverGroupLaunchConfigurations:GetAppLaunchConfigurationResponse', getAppLaunchConfigurationResponse_serverGroupLaunchConfigurations - The launch configurations for server groups in this application.

$sel:autoLaunch:GetAppLaunchConfigurationResponse', getAppLaunchConfigurationResponse_autoLaunch - Indicates whether the application is configured to launch automatically after replication is complete.

$sel:roleName:GetAppLaunchConfigurationResponse', getAppLaunchConfigurationResponse_roleName - The name of the service role in the customer's account that AWS CloudFormation uses to launch the application.

$sel:appId:GetAppLaunchConfiguration', getAppLaunchConfigurationResponse_appId - The ID of the application.

$sel:httpStatus:GetAppLaunchConfigurationResponse', getAppLaunchConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getAppLaunchConfigurationResponse_serverGroupLaunchConfigurations :: Lens' GetAppLaunchConfigurationResponse (Maybe [ServerGroupLaunchConfiguration])

The launch configurations for server groups in this application.

getAppLaunchConfigurationResponse_autoLaunch :: Lens' GetAppLaunchConfigurationResponse (Maybe Bool)

Indicates whether the application is configured to launch automatically after replication is complete.

getAppLaunchConfigurationResponse_roleName :: Lens' GetAppLaunchConfigurationResponse (Maybe Text)

The name of the service role in the customer's account that AWS CloudFormation uses to launch the application.

getAppLaunchConfigurationResponse_appId :: Lens' GetAppLaunchConfigurationResponse (Maybe Text)

The ID of the application.

getAppLaunchConfigurationResponse_httpStatus :: Lens' GetAppLaunchConfigurationResponse Int

The response's http status code.