Amazonka.SMS.Types.Connector

Description

Documentation

data Connector

Represents a connector.

See: newConnector smart constructor.

Constructors

Connector'

Fields

status :: Maybe ConnectorStatus

The status of the connector.

vmManagerName :: Maybe Text

The name of the VM manager.

ipAddress :: Maybe Text

The IP address of the connector.

vmManagerId :: Maybe Text

The ID of the VM manager.

vmManagerType :: Maybe VmManagerType

The VM management product.

connectorId :: Maybe Text

The ID of the connector.

associatedOn :: Maybe POSIX

The time the connector was associated.

macAddress :: Maybe Text

The MAC address of the connector.

version :: Maybe Text

The connector version.

capabilityList :: Maybe [ConnectorCapability]

The capabilities of the connector.

newConnector :: Connector

Create a value of Connector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Connector', connector_status - The status of the connector.

$sel:vmManagerName:Connector', connector_vmManagerName - The name of the VM manager.

$sel:ipAddress:Connector', connector_ipAddress - The IP address of the connector.

$sel:vmManagerId:Connector', connector_vmManagerId - The ID of the VM manager.

$sel:vmManagerType:Connector', connector_vmManagerType - The VM management product.

$sel:connectorId:Connector', connector_connectorId - The ID of the connector.

$sel:associatedOn:Connector', connector_associatedOn - The time the connector was associated.

$sel:macAddress:Connector', connector_macAddress - The MAC address of the connector.

$sel:version:Connector', connector_version - The connector version.

$sel:capabilityList:Connector', connector_capabilityList - The capabilities of the connector.

connector_status :: Lens' Connector (Maybe ConnectorStatus)

The status of the connector.

connector_vmManagerName :: Lens' Connector (Maybe Text)

The name of the VM manager.

connector_ipAddress :: Lens' Connector (Maybe Text)

The IP address of the connector.

connector_vmManagerId :: Lens' Connector (Maybe Text)

The ID of the VM manager.

connector_vmManagerType :: Lens' Connector (Maybe VmManagerType)

The VM management product.

connector_connectorId :: Lens' Connector (Maybe Text)

The ID of the connector.

connector_associatedOn :: Lens' Connector (Maybe UTCTime)

The time the connector was associated.

connector_macAddress :: Lens' Connector (Maybe Text)

The MAC address of the connector.

connector_version :: Lens' Connector (Maybe Text)

The connector version.

connector_capabilityList :: Lens' Connector (Maybe [ConnectorCapability])

The capabilities of the connector.