Amazonka.SMS.Types.ServerLaunchConfiguration

Description

Documentation

data ServerLaunchConfiguration

Launch configuration for a server.

See: newServerLaunchConfiguration smart constructor.

Constructors

ServerLaunchConfiguration'

Fields

ec2KeyName :: Maybe Text

The name of the Amazon EC2 SSH key to be used for connecting to the launched server.

configureScriptType :: Maybe ScriptType

The type of configuration script.

associatePublicIpAddress :: Maybe Bool

Indicates whether a publicly accessible IP address is created when launching the server.

iamInstanceProfileName :: Maybe Text

The name of the IAM instance profile.

subnet :: Maybe Text

The ID of the subnet the server should be launched into.

logicalId :: Maybe Text

The logical ID of the server in the AWS CloudFormation template.

securityGroup :: Maybe Text

The ID of the security group that applies to the launched server.

userData :: Maybe UserData

Location of the user-data script to be executed when launching the server.

instanceType :: Maybe Text

The instance type to use when launching the server.

configureScript :: Maybe S3Location

server :: Maybe Server

The ID of the server with which the launch configuration is associated.

vpc :: Maybe Text

The ID of the VPC into which the server should be launched.

newServerLaunchConfiguration :: ServerLaunchConfiguration

Create a value of ServerLaunchConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ec2KeyName:ServerLaunchConfiguration', serverLaunchConfiguration_ec2KeyName - The name of the Amazon EC2 SSH key to be used for connecting to the launched server.

$sel:configureScriptType:ServerLaunchConfiguration', serverLaunchConfiguration_configureScriptType - The type of configuration script.

$sel:associatePublicIpAddress:ServerLaunchConfiguration', serverLaunchConfiguration_associatePublicIpAddress - Indicates whether a publicly accessible IP address is created when launching the server.

$sel:iamInstanceProfileName:ServerLaunchConfiguration', serverLaunchConfiguration_iamInstanceProfileName - The name of the IAM instance profile.

$sel:subnet:ServerLaunchConfiguration', serverLaunchConfiguration_subnet - The ID of the subnet the server should be launched into.

$sel:logicalId:ServerLaunchConfiguration', serverLaunchConfiguration_logicalId - The logical ID of the server in the AWS CloudFormation template.

$sel:securityGroup:ServerLaunchConfiguration', serverLaunchConfiguration_securityGroup - The ID of the security group that applies to the launched server.

$sel:userData:ServerLaunchConfiguration', serverLaunchConfiguration_userData - Location of the user-data script to be executed when launching the server.

$sel:instanceType:ServerLaunchConfiguration', serverLaunchConfiguration_instanceType - The instance type to use when launching the server.

$sel:configureScript:ServerLaunchConfiguration', serverLaunchConfiguration_configureScript - Undocumented member.

$sel:server:ServerLaunchConfiguration', serverLaunchConfiguration_server - The ID of the server with which the launch configuration is associated.

$sel:vpc:ServerLaunchConfiguration', serverLaunchConfiguration_vpc - The ID of the VPC into which the server should be launched.

serverLaunchConfiguration_ec2KeyName :: Lens' ServerLaunchConfiguration (Maybe Text)

The name of the Amazon EC2 SSH key to be used for connecting to the launched server.

serverLaunchConfiguration_configureScriptType :: Lens' ServerLaunchConfiguration (Maybe ScriptType)

The type of configuration script.

serverLaunchConfiguration_associatePublicIpAddress :: Lens' ServerLaunchConfiguration (Maybe Bool)

Indicates whether a publicly accessible IP address is created when launching the server.

serverLaunchConfiguration_iamInstanceProfileName :: Lens' ServerLaunchConfiguration (Maybe Text)

The name of the IAM instance profile.

serverLaunchConfiguration_subnet :: Lens' ServerLaunchConfiguration (Maybe Text)

The ID of the subnet the server should be launched into.

serverLaunchConfiguration_logicalId :: Lens' ServerLaunchConfiguration (Maybe Text)

The logical ID of the server in the AWS CloudFormation template.

serverLaunchConfiguration_securityGroup :: Lens' ServerLaunchConfiguration (Maybe Text)

The ID of the security group that applies to the launched server.

serverLaunchConfiguration_userData :: Lens' ServerLaunchConfiguration (Maybe UserData)

Location of the user-data script to be executed when launching the server.

serverLaunchConfiguration_instanceType :: Lens' ServerLaunchConfiguration (Maybe Text)

The instance type to use when launching the server.

serverLaunchConfiguration_configureScript :: Lens' ServerLaunchConfiguration (Maybe S3Location)

Undocumented member.

serverLaunchConfiguration_server :: Lens' ServerLaunchConfiguration (Maybe Server)

The ID of the server with which the launch configuration is associated.

serverLaunchConfiguration_vpc :: Lens' ServerLaunchConfiguration (Maybe Text)

The ID of the VPC into which the server should be launched.