Amazonka.SNS.CreatePlatformApplication

Description

Creates a platform application object for one of the supported push notification services, such as APNS and GCM (Firebase Cloud Messaging), to which devices and mobile apps may register. You must specify PlatformPrincipal and PlatformCredential attributes when using the CreatePlatformApplication action.

PlatformPrincipal and PlatformCredential are received from the notification service.

For ADM, PlatformPrincipal is client id and PlatformCredential is client secret.

For Baidu, PlatformPrincipal is API key and PlatformCredential is secret key.

For APNS and APNS_SANDBOX, PlatformPrincipal is SSL certificate and PlatformCredential is private key.

For GCM (Firebase Cloud Messaging), there is no PlatformPrincipal and the PlatformCredential is API key.

For MPNS, PlatformPrincipal is TLS certificate and PlatformCredential is private key.

For WNS, PlatformPrincipal is Package Security Identifier and PlatformCredential is secret key.

You can use the returned PlatformApplicationArn as an attribute for the CreatePlatformEndpoint action.

Creating a Request

data CreatePlatformApplication

Input for CreatePlatformApplication action.

See: newCreatePlatformApplication smart constructor.

Constructors

CreatePlatformApplication'

Fields

name :: Text

Application names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, hyphens, and periods, and must be between 1 and 256 characters long.

platform :: Text

The following platforms are supported: ADM (Amazon Device Messaging), APNS (Apple Push Notification Service), APNS_SANDBOX, and GCM (Firebase Cloud Messaging).

attributes :: HashMap Text Text

For a list of attributes, see SetPlatformApplicationAttributes.

newCreatePlatformApplication

Arguments

:: Text
$sel:name:CreatePlatformApplication'

-> Text
$sel:platform:CreatePlatformApplication'

-> CreatePlatformApplication

Create a value of CreatePlatformApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreatePlatformApplication', createPlatformApplication_name - Application names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, hyphens, and periods, and must be between 1 and 256 characters long.

$sel:platform:CreatePlatformApplication', createPlatformApplication_platform - The following platforms are supported: ADM (Amazon Device Messaging), APNS (Apple Push Notification Service), APNS_SANDBOX, and GCM (Firebase Cloud Messaging).

$sel:attributes:CreatePlatformApplication', createPlatformApplication_attributes - For a list of attributes, see SetPlatformApplicationAttributes.

Request Lenses

createPlatformApplication_name :: Lens' CreatePlatformApplication Text

Application names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, hyphens, and periods, and must be between 1 and 256 characters long.

createPlatformApplication_platform :: Lens' CreatePlatformApplication Text

The following platforms are supported: ADM (Amazon Device Messaging), APNS (Apple Push Notification Service), APNS_SANDBOX, and GCM (Firebase Cloud Messaging).

createPlatformApplication_attributes :: Lens' CreatePlatformApplication (HashMap Text Text)

For a list of attributes, see SetPlatformApplicationAttributes.

Destructuring the Response

data CreatePlatformApplicationResponse

Response from CreatePlatformApplication action.

See: newCreatePlatformApplicationResponse smart constructor.

Constructors

CreatePlatformApplicationResponse'

Fields

platformApplicationArn :: Maybe Text

PlatformApplicationArn is returned.

httpStatus :: Int

The response's http status code.

newCreatePlatformApplicationResponse

Arguments

:: Int
$sel:httpStatus:CreatePlatformApplicationResponse'

-> CreatePlatformApplicationResponse

Create a value of CreatePlatformApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platformApplicationArn:CreatePlatformApplicationResponse', createPlatformApplicationResponse_platformApplicationArn - PlatformApplicationArn is returned.

$sel:httpStatus:CreatePlatformApplicationResponse', createPlatformApplicationResponse_httpStatus - The response's http status code.

Response Lenses

createPlatformApplicationResponse_platformApplicationArn :: Lens' CreatePlatformApplicationResponse (Maybe Text)

PlatformApplicationArn is returned.

createPlatformApplicationResponse_httpStatus :: Lens' CreatePlatformApplicationResponse Int

The response's http status code.