Amazonka.SNS.Lens

Description

Operations

ListPhoneNumbersOptedOut

listPhoneNumbersOptedOut_nextToken :: Lens' ListPhoneNumbersOptedOut (Maybe Text)

A NextToken string is used when you call the ListPhoneNumbersOptedOut action to retrieve additional records that are available after the first page of results.

listPhoneNumbersOptedOutResponse_phoneNumbers :: Lens' ListPhoneNumbersOptedOutResponse (Maybe [Text])

A list of phone numbers that are opted out of receiving SMS messages. The list is paginated, and each page can contain up to 100 phone numbers.

listPhoneNumbersOptedOutResponse_nextToken :: Lens' ListPhoneNumbersOptedOutResponse (Maybe Text)

A NextToken string is returned when you call the ListPhoneNumbersOptedOut action if additional records are available after the first page of results.

listPhoneNumbersOptedOutResponse_httpStatus :: Lens' ListPhoneNumbersOptedOutResponse Int

The response's http status code.

DeleteEndpoint

deleteEndpoint_endpointArn :: Lens' DeleteEndpoint Text

EndpointArn of endpoint to delete.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the topic for which to list tags.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

The tags associated with the specified topic.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

RemovePermission

removePermission_topicArn :: Lens' RemovePermission Text

The ARN of the topic whose access control policy you wish to modify.

removePermission_label :: Lens' RemovePermission Text

The unique label of the statement you want to remove.

DeleteTopic

deleteTopic_topicArn :: Lens' DeleteTopic Text

The ARN of the topic you want to delete.

SetSMSAttributes

setSMSAttributes_attributes :: Lens' SetSMSAttributes (HashMap Text Text)

The default settings for sending SMS messages from your account. You can set values for the following attribute names:

MonthlySpendLimit – The maximum amount in USD that you are willing to spend each month to send SMS messages. When Amazon SNS determines that sending an SMS message would incur a cost that exceeds this limit, it stops sending SMS messages within minutes.

Amazon SNS stops sending SMS messages within minutes of the limit being crossed. During that interval, if you continue to send SMS messages, you will incur costs that exceed your limit.

By default, the spend limit is set to the maximum allowed by Amazon SNS. If you want to raise the limit, submit an SNS Limit Increase case. For New limit value, enter your desired monthly spend limit. In the Use Case Description field, explain that you are requesting an SMS monthly spend limit increase.

DeliveryStatusIAMRole – The ARN of the IAM role that allows Amazon SNS to write logs about SMS deliveries in CloudWatch Logs. For each SMS message that you send, Amazon SNS writes a log that includes the message price, the success or failure status, the reason for failure (if the message failed), the message dwell time, and other information.

DeliveryStatusSuccessSamplingRate – The percentage of successful SMS deliveries for which Amazon SNS will write logs in CloudWatch Logs. The value can be an integer from 0 - 100. For example, to write logs only for failed deliveries, set this value to 0. To write logs for 10% of your successful deliveries, set it to 10.

DefaultSenderID – A string, such as your business brand, that is displayed as the sender on the receiving device. Support for sender IDs varies by country. The sender ID can be 1 - 11 alphanumeric characters, and it must contain at least one letter.

DefaultSMSType – The type of SMS message that you will send by default. You can assign the following values:

Promotional – (Default) Noncritical messages, such as marketing messages. Amazon SNS optimizes the message delivery to incur the lowest cost.

Transactional – Critical messages that support customer transactions, such as one-time passcodes for multi-factor authentication. Amazon SNS optimizes the message delivery to achieve the highest reliability.

UsageReportS3Bucket – The name of the Amazon S3 bucket to receive daily SMS usage reports from Amazon SNS. Each day, Amazon SNS will deliver a usage report as a CSV file to the bucket. The report includes the following information for each SMS message that was successfully delivered by your account:

Time that the message was published (in UTC)

Message ID

Destination phone number

Message type

Delivery status

Message price (in USD)

Part number (a message is split into multiple parts if it is too long for a single message)

Total number of parts

To receive the report, the bucket must have a policy that allows the Amazon SNS service principal to perform the s3:PutObject and s3:GetBucketLocation actions.

For an example bucket policy and usage report, see Monitoring SMS Activity in the Amazon SNS Developer Guide.

setSMSAttributesResponse_httpStatus :: Lens' SetSMSAttributesResponse Int

The response's http status code.

ListTopics

listTopics_nextToken :: Lens' ListTopics (Maybe Text)

Token returned by the previous ListTopics request.

listTopicsResponse_topics :: Lens' ListTopicsResponse (Maybe [Topic])

A list of topic ARNs.

listTopicsResponse_nextToken :: Lens' ListTopicsResponse (Maybe Text)

Token to pass along to the next ListTopics request. This element is returned if there are additional topics to retrieve.

listTopicsResponse_httpStatus :: Lens' ListTopicsResponse Int

The response's http status code.

VerifySMSSandboxPhoneNumber

verifySMSSandboxPhoneNumber_phoneNumber :: Lens' VerifySMSSandboxPhoneNumber Text

The destination phone number to verify.

verifySMSSandboxPhoneNumber_oneTimePassword :: Lens' VerifySMSSandboxPhoneNumber Text

The OTP sent to the destination number from the CreateSMSSandBoxPhoneNumber call.

verifySMSSandboxPhoneNumberResponse_httpStatus :: Lens' VerifySMSSandboxPhoneNumberResponse Int

The response's http status code.

CreatePlatformEndpoint

createPlatformEndpoint_customUserData :: Lens' CreatePlatformEndpoint (Maybe Text)

Arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

createPlatformEndpoint_attributes :: Lens' CreatePlatformEndpoint (Maybe (HashMap Text Text))

For a list of attributes, see SetEndpointAttributes.

createPlatformEndpoint_platformApplicationArn :: Lens' CreatePlatformEndpoint Text

PlatformApplicationArn returned from CreatePlatformApplication is used to create a an endpoint.

createPlatformEndpoint_token :: Lens' CreatePlatformEndpoint Text

Unique identifier created by the notification service for an app on a device. The specific name for Token will vary, depending on which notification service is being used. For example, when using APNS as the notification service, you need the device token. Alternatively, when using GCM (Firebase Cloud Messaging) or ADM, the device token equivalent is called the registration ID.

createPlatformEndpointResponse_endpointArn :: Lens' CreatePlatformEndpointResponse (Maybe Text)

EndpointArn returned from CreateEndpoint action.

createPlatformEndpointResponse_httpStatus :: Lens' CreatePlatformEndpointResponse Int

The response's http status code.

SetPlatformApplicationAttributes

setPlatformApplicationAttributes_platformApplicationArn :: Lens' SetPlatformApplicationAttributes Text

PlatformApplicationArn for SetPlatformApplicationAttributes action.

setPlatformApplicationAttributes_attributes :: Lens' SetPlatformApplicationAttributes (HashMap Text Text)

A map of the platform application attributes. Attributes in this map include the following:

PlatformCredential – The credential received from the notification service. For APNS and APNS_SANDBOX, PlatformCredential is private key. For GCM (Firebase Cloud Messaging), PlatformCredential is API key. For ADM, PlatformCredential is client secret.

PlatformPrincipal – The principal received from the notification service. For APNS and APNS_SANDBOX, PlatformPrincipal is SSL certificate. For GCM (Firebase Cloud Messaging), there is no PlatformPrincipal. For ADM, PlatformPrincipal is client id.

EventEndpointCreated – Topic ARN to which EndpointCreated event notifications are sent.

EventEndpointDeleted – Topic ARN to which EndpointDeleted event notifications are sent.

EventEndpointUpdated – Topic ARN to which EndpointUpdate event notifications are sent.

EventDeliveryFailure – Topic ARN to which DeliveryFailure event notifications are sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.

SuccessFeedbackRoleArn – IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf.

FailureFeedbackRoleArn – IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf.

SuccessFeedbackSampleRate – Sample rate percentage (0-100) of successfully delivered messages.

ListSubscriptionsByTopic

listSubscriptionsByTopic_nextToken :: Lens' ListSubscriptionsByTopic (Maybe Text)

Token returned by the previous ListSubscriptionsByTopic request.

listSubscriptionsByTopic_topicArn :: Lens' ListSubscriptionsByTopic Text

The ARN of the topic for which you wish to find subscriptions.

listSubscriptionsByTopicResponse_nextToken :: Lens' ListSubscriptionsByTopicResponse (Maybe Text)

Token to pass along to the next ListSubscriptionsByTopic request. This element is returned if there are more subscriptions to retrieve.

listSubscriptionsByTopicResponse_subscriptions :: Lens' ListSubscriptionsByTopicResponse (Maybe [Subscription])

A list of subscriptions.

listSubscriptionsByTopicResponse_httpStatus :: Lens' ListSubscriptionsByTopicResponse Int

The response's http status code.

GetTopicAttributes

getTopicAttributes_topicArn :: Lens' GetTopicAttributes Text

The ARN of the topic whose properties you want to get.

getTopicAttributesResponse_attributes :: Lens' GetTopicAttributesResponse (Maybe (HashMap Text Text))

A map of the topic's attributes. Attributes in this map include the following:

DeliveryPolicy – The JSON serialization of the topic's delivery policy.

DisplayName – The human-readable name used in the From field for notifications to email and email-json endpoints.

Owner – The account ID of the topic's owner.

Policy – The JSON serialization of the topic's access control policy.

SubscriptionsConfirmed – The number of confirmed subscriptions for the topic.

SubscriptionsDeleted – The number of deleted subscriptions for the topic.

SubscriptionsPending – The number of subscriptions pending confirmation for the topic.

TopicArn – The topic's ARN.

EffectiveDeliveryPolicy – The JSON serialization of the effective delivery policy, taking system defaults into account.

The following attribute applies only to server-side-encryption:

KmsMasterKeyId - The ID of an Amazon Web Services managed customer master key (CMK) for Amazon SNS or a custom CMK. For more information, see Key Terms. For more examples, see KeyId in the Key Management Service API Reference.

The following attributes apply only to FIFO topics:

FifoTopic – When this is set to true, a FIFO topic is created.

ContentBasedDeduplication – Enables content-based deduplication for FIFO topics.

By default, ContentBasedDeduplication is set to false. If you create a FIFO topic and this attribute is false, you must specify a value for the MessageDeduplicationId parameter for the Publish action.

When you set ContentBasedDeduplication to true, Amazon SNS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

(Optional) To override the generated value, you can specify a value for the MessageDeduplicationId parameter for the Publish action.

getTopicAttributesResponse_httpStatus :: Lens' GetTopicAttributesResponse Int

The response's http status code.

CreateSMSSandboxPhoneNumber

createSMSSandboxPhoneNumber_languageCode :: Lens' CreateSMSSandboxPhoneNumber (Maybe LanguageCodeString)

The language to use for sending the OTP. The default value is en-US.

createSMSSandboxPhoneNumber_phoneNumber :: Lens' CreateSMSSandboxPhoneNumber Text

The destination phone number to verify. On verification, Amazon SNS adds this phone number to the list of verified phone numbers that you can send SMS messages to.

createSMSSandboxPhoneNumberResponse_httpStatus :: Lens' CreateSMSSandboxPhoneNumberResponse Int

The response's http status code.

OptInPhoneNumber

optInPhoneNumber_phoneNumber :: Lens' OptInPhoneNumber Text

The phone number to opt in. Use E.164 format.

optInPhoneNumberResponse_httpStatus :: Lens' OptInPhoneNumberResponse Int

The response's http status code.

DeleteSMSSandboxPhoneNumber

deleteSMSSandboxPhoneNumber_phoneNumber :: Lens' DeleteSMSSandboxPhoneNumber Text

The destination phone number to delete.

deleteSMSSandboxPhoneNumberResponse_httpStatus :: Lens' DeleteSMSSandboxPhoneNumberResponse Int

The response's http status code.

ListSMSSandboxPhoneNumbers

listSMSSandboxPhoneNumbers_nextToken :: Lens' ListSMSSandboxPhoneNumbers (Maybe Text)

Token that the previous ListSMSSandboxPhoneNumbersInput request returns.

listSMSSandboxPhoneNumbers_maxResults :: Lens' ListSMSSandboxPhoneNumbers (Maybe Natural)

The maximum number of phone numbers to return.

listSMSSandboxPhoneNumbersResponse_nextToken :: Lens' ListSMSSandboxPhoneNumbersResponse (Maybe Text)

A NextToken string is returned when you call the ListSMSSandboxPhoneNumbersInput operation if additional pages of records are available.

listSMSSandboxPhoneNumbersResponse_httpStatus :: Lens' ListSMSSandboxPhoneNumbersResponse Int

The response's http status code.

listSMSSandboxPhoneNumbersResponse_phoneNumbers :: Lens' ListSMSSandboxPhoneNumbersResponse [SMSSandboxPhoneNumber]

A list of the calling account's pending and verified phone numbers.

CreatePlatformApplication

createPlatformApplication_name :: Lens' CreatePlatformApplication Text

Application names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, hyphens, and periods, and must be between 1 and 256 characters long.

createPlatformApplication_platform :: Lens' CreatePlatformApplication Text

The following platforms are supported: ADM (Amazon Device Messaging), APNS (Apple Push Notification Service), APNS_SANDBOX, and GCM (Firebase Cloud Messaging).

createPlatformApplication_attributes :: Lens' CreatePlatformApplication (HashMap Text Text)

For a list of attributes, see SetPlatformApplicationAttributes.

createPlatformApplicationResponse_platformApplicationArn :: Lens' CreatePlatformApplicationResponse (Maybe Text)

PlatformApplicationArn is returned.

createPlatformApplicationResponse_httpStatus :: Lens' CreatePlatformApplicationResponse Int

The response's http status code.

GetPlatformApplicationAttributes

getPlatformApplicationAttributes_platformApplicationArn :: Lens' GetPlatformApplicationAttributes Text

PlatformApplicationArn for GetPlatformApplicationAttributesInput.

getPlatformApplicationAttributesResponse_attributes :: Lens' GetPlatformApplicationAttributesResponse (Maybe (HashMap Text Text))

Attributes include the following:

EventEndpointCreated – Topic ARN to which EndpointCreated event notifications should be sent.

EventEndpointDeleted – Topic ARN to which EndpointDeleted event notifications should be sent.

EventEndpointUpdated – Topic ARN to which EndpointUpdate event notifications should be sent.

EventDeliveryFailure – Topic ARN to which DeliveryFailure event notifications should be sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.

getPlatformApplicationAttributesResponse_httpStatus :: Lens' GetPlatformApplicationAttributesResponse Int

The response's http status code.

ListEndpointsByPlatformApplication

listEndpointsByPlatformApplication_nextToken :: Lens' ListEndpointsByPlatformApplication (Maybe Text)

NextToken string is used when calling ListEndpointsByPlatformApplication action to retrieve additional records that are available after the first page results.

listEndpointsByPlatformApplication_platformApplicationArn :: Lens' ListEndpointsByPlatformApplication Text

PlatformApplicationArn for ListEndpointsByPlatformApplicationInput action.

listEndpointsByPlatformApplicationResponse_nextToken :: Lens' ListEndpointsByPlatformApplicationResponse (Maybe Text)

NextToken string is returned when calling ListEndpointsByPlatformApplication action if additional records are available after the first page results.

listEndpointsByPlatformApplicationResponse_endpoints :: Lens' ListEndpointsByPlatformApplicationResponse (Maybe [Endpoint])

Endpoints returned for ListEndpointsByPlatformApplication action.

listEndpointsByPlatformApplicationResponse_httpStatus :: Lens' ListEndpointsByPlatformApplicationResponse Int

The response's http status code.

SetTopicAttributes

setTopicAttributes_attributeValue :: Lens' SetTopicAttributes (Maybe Text)

The new value for the attribute.

setTopicAttributes_topicArn :: Lens' SetTopicAttributes Text

The ARN of the topic to modify.

setTopicAttributes_attributeName :: Lens' SetTopicAttributes Text

A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that the SetTopicAttributes action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

DisplayName – The display name to use for a topic with SMS subscriptions.

Policy – The policy that defines who can access your topic. By default, only the topic owner can publish or subscribe to the topic.

The following attribute applies only to server-side-encryption:

KmsMasterKeyId – The ID of an Amazon Web Services managed customer master key (CMK) for Amazon SNS or a custom CMK. For more information, see Key Terms. For more examples, see KeyId in the Key Management Service API Reference.

The following attribute applies only to FIFO topics:

ContentBasedDeduplication – Enables content-based deduplication for FIFO topics.

By default, ContentBasedDeduplication is set to false. If you create a FIFO topic and this attribute is false, you must specify a value for the MessageDeduplicationId parameter for the Publish action.

When you set ContentBasedDeduplication to true, Amazon SNS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

(Optional) To override the generated value, you can specify a value for the MessageDeduplicationId parameter for the Publish action.

DeletePlatformApplication

deletePlatformApplication_platformApplicationArn :: Lens' DeletePlatformApplication Text

PlatformApplicationArn of platform application object to delete.

GetSMSAttributes

getSMSAttributes_attributes :: Lens' GetSMSAttributes (Maybe [Text])

A list of the individual attribute names, such as MonthlySpendLimit, for which you want values.

For all attribute names, see SetSMSAttributes.

If you don't use this parameter, Amazon SNS returns all SMS attributes.

getSMSAttributesResponse_attributes :: Lens' GetSMSAttributesResponse (Maybe (HashMap Text Text))

The SMS attribute names and their values.

getSMSAttributesResponse_httpStatus :: Lens' GetSMSAttributesResponse Int

The response's http status code.

ListPlatformApplications

listPlatformApplications_nextToken :: Lens' ListPlatformApplications (Maybe Text)

NextToken string is used when calling ListPlatformApplications action to retrieve additional records that are available after the first page results.

listPlatformApplicationsResponse_platformApplications :: Lens' ListPlatformApplicationsResponse (Maybe [PlatformApplication])

Platform applications returned when calling ListPlatformApplications action.

listPlatformApplicationsResponse_nextToken :: Lens' ListPlatformApplicationsResponse (Maybe Text)

NextToken string is returned when calling ListPlatformApplications action if additional records are available after the first page results.

listPlatformApplicationsResponse_httpStatus :: Lens' ListPlatformApplicationsResponse Int

The response's http status code.

AddPermission

addPermission_topicArn :: Lens' AddPermission Text

The ARN of the topic whose access control policy you wish to modify.

addPermission_label :: Lens' AddPermission Text

A unique identifier for the new policy statement.

addPermission_aWSAccountId :: Lens' AddPermission [Text]

The account IDs of the users (principals) who will be given access to the specified actions. The users must have account, but do not need to be signed up for this service.

addPermission_actionName :: Lens' AddPermission [Text]

The action you want to allow for the specified principal(s).

Valid values: Any Amazon SNS action name, for example Publish.

GetEndpointAttributes

getEndpointAttributes_endpointArn :: Lens' GetEndpointAttributes Text

EndpointArn for GetEndpointAttributes input.

getEndpointAttributesResponse_attributes :: Lens' GetEndpointAttributesResponse (Maybe (HashMap Text Text))

Attributes include the following:

CustomUserData – arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

Enabled – flag that enables/disables delivery to the endpoint. Amazon SNS will set this to false when a notification service indicates to Amazon SNS that the endpoint is invalid. Users can set it back to true, typically after updating Token.

Token – device token, also referred to as a registration id, for an app and mobile device. This is returned from the notification service when an app and mobile device are registered with the notification service.

The device token for the iOS platform is returned in lowercase.

getEndpointAttributesResponse_httpStatus :: Lens' GetEndpointAttributesResponse Int

The response's http status code.

ListSubscriptions

listSubscriptions_nextToken :: Lens' ListSubscriptions (Maybe Text)

Token returned by the previous ListSubscriptions request.

listSubscriptionsResponse_nextToken :: Lens' ListSubscriptionsResponse (Maybe Text)

Token to pass along to the next ListSubscriptions request. This element is returned if there are more subscriptions to retrieve.

listSubscriptionsResponse_subscriptions :: Lens' ListSubscriptionsResponse (Maybe [Subscription])

A list of subscriptions.

listSubscriptionsResponse_httpStatus :: Lens' ListSubscriptionsResponse Int

The response's http status code.

GetSubscriptionAttributes

getSubscriptionAttributes_subscriptionArn :: Lens' GetSubscriptionAttributes Text

The ARN of the subscription whose properties you want to get.

getSubscriptionAttributesResponse_attributes :: Lens' GetSubscriptionAttributesResponse (Maybe (HashMap Text Text))

A map of the subscription's attributes. Attributes in this map include the following:

ConfirmationWasAuthenticated – true if the subscription confirmation request was authenticated.

DeliveryPolicy – The JSON serialization of the subscription's delivery policy.

EffectiveDeliveryPolicy – The JSON serialization of the effective delivery policy that takes into account the topic delivery policy and account system defaults.

FilterPolicy – The filter policy JSON that is assigned to the subscription. For more information, see Amazon SNS Message Filtering in the Amazon SNS Developer Guide.

Owner – The account ID of the subscription's owner.

PendingConfirmation – true if the subscription hasn't been confirmed. To confirm a pending subscription, call the ConfirmSubscription action with a confirmation token.

RawMessageDelivery – true if raw message delivery is enabled for the subscription. Raw messages are free of JSON formatting and can be sent to HTTP/S and Amazon SQS endpoints.

RedrivePolicy – When specified, sends undeliverable messages to the specified Amazon SQS dead-letter queue. Messages that can't be delivered due to client errors (for example, when the subscribed endpoint is unreachable) or server errors (for example, when the service that powers the subscribed endpoint becomes unavailable) are held in the dead-letter queue for further analysis or reprocessing.

SubscriptionArn – The subscription's ARN.

TopicArn – The topic ARN that the subscription is associated with.

The following attribute applies only to Amazon Kinesis Data Firehose delivery stream subscriptions:

SubscriptionRoleArn – The ARN of the IAM role that has the following:

Permission to write to the Kinesis Data Firehose delivery stream

Amazon SNS listed as a trusted entity

Specifying a valid ARN for this attribute is required for Kinesis Data Firehose delivery stream subscriptions. For more information, see Fanout to Kinesis Data Firehose delivery streams in the Amazon SNS Developer Guide.

getSubscriptionAttributesResponse_httpStatus :: Lens' GetSubscriptionAttributesResponse Int

The response's http status code.

CreateTopic

createTopic_attributes :: Lens' CreateTopic (Maybe (HashMap Text Text))

A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that the CreateTopic action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

DisplayName – The display name to use for a topic with SMS subscriptions.

FifoTopic – Set to true to create a FIFO topic.

Policy – The policy that defines who can access your topic. By default, only the topic owner can publish or subscribe to the topic.

The following attribute applies only to server-side encryption:

KmsMasterKeyId – The ID of an Amazon Web Services managed customer master key (CMK) for Amazon SNS or a custom CMK. For more information, see Key Terms. For more examples, see KeyId in the Key Management Service API Reference.

The following attributes apply only to FIFO topics:

FifoTopic – When this is set to true, a FIFO topic is created.

ContentBasedDeduplication – Enables content-based deduplication for FIFO topics.

By default, ContentBasedDeduplication is set to false. If you create a FIFO topic and this attribute is false, you must specify a value for the MessageDeduplicationId parameter for the Publish action.

When you set ContentBasedDeduplication to true, Amazon SNS uses a SHA-256 hash to generate the MessageDeduplicationId using the body of the message (but not the attributes of the message).

(Optional) To override the generated value, you can specify a value for the MessageDeduplicationId parameter for the Publish action.

createTopic_tags :: Lens' CreateTopic (Maybe [Tag])

The list of tags to add to a new topic.

To be able to tag a topic on creation, you must have the sns:CreateTopic and sns:TagResource permissions.

createTopic_name :: Lens' CreateTopic Text

The name of the topic you want to create.

Constraints: Topic names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, and hyphens, and must be between 1 and 256 characters long.

For a FIFO (first-in-first-out) topic, the name must end with the .fifo suffix.

createTopicResponse_topicArn :: Lens' CreateTopicResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the created topic.

createTopicResponse_httpStatus :: Lens' CreateTopicResponse Int

The response's http status code.

CheckIfPhoneNumberIsOptedOut

checkIfPhoneNumberIsOptedOut_phoneNumber :: Lens' CheckIfPhoneNumberIsOptedOut Text

The phone number for which you want to check the opt out status.

checkIfPhoneNumberIsOptedOutResponse_isOptedOut :: Lens' CheckIfPhoneNumberIsOptedOutResponse (Maybe Bool)

Indicates whether the phone number is opted out:

true – The phone number is opted out, meaning you cannot publish SMS messages to it.

false – The phone number is opted in, meaning you can publish SMS messages to it.

checkIfPhoneNumberIsOptedOutResponse_httpStatus :: Lens' CheckIfPhoneNumberIsOptedOutResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the topic to which to add tags.

tagResource_tags :: Lens' TagResource [Tag]

The tags to be added to the specified topic. A tag consists of a required key and an optional value.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

Subscribe

subscribe_returnSubscriptionArn :: Lens' Subscribe (Maybe Bool)

Sets whether the response from the Subscribe request includes the subscription ARN, even if the subscription is not yet confirmed.

If you set this parameter to true, the response includes the ARN in all cases, even if the subscription is not yet confirmed. In addition to the ARN for confirmed subscriptions, the response also includes the pending subscription ARN value for subscriptions that aren't yet confirmed. A subscription becomes confirmed when the subscriber calls the ConfirmSubscription action with a confirmation token.

The default value is false.

subscribe_attributes :: Lens' Subscribe (Maybe (HashMap Text Text))

A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that the Subscribe action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

FilterPolicy – The simple JSON object that lets your subscriber receive only a subset of messages, rather than receiving every message published to the topic.

RawMessageDelivery – When set to true, enables raw message delivery to Amazon SQS or HTTP/S endpoints. This eliminates the need for the endpoints to process JSON formatting, which is otherwise created for Amazon SNS metadata.

RedrivePolicy – When specified, sends undeliverable messages to the specified Amazon SQS dead-letter queue. Messages that can't be delivered due to client errors (for example, when the subscribed endpoint is unreachable) or server errors (for example, when the service that powers the subscribed endpoint becomes unavailable) are held in the dead-letter queue for further analysis or reprocessing.

The following attribute applies only to Amazon Kinesis Data Firehose delivery stream subscriptions:

SubscriptionRoleArn – The ARN of the IAM role that has the following:

Permission to write to the Kinesis Data Firehose delivery stream

Amazon SNS listed as a trusted entity

Specifying a valid ARN for this attribute is required for Kinesis Data Firehose delivery stream subscriptions. For more information, see Fanout to Kinesis Data Firehose delivery streams in the Amazon SNS Developer Guide.

subscribe_endpoint :: Lens' Subscribe (Maybe Text)

The endpoint that you want to receive notifications. Endpoints vary by protocol:

For the http protocol, the (public) endpoint is a URL beginning with http://.

For the https protocol, the (public) endpoint is a URL beginning with https://.

For the email protocol, the endpoint is an email address.

For the email-json protocol, the endpoint is an email address.

For the sms protocol, the endpoint is a phone number of an SMS-enabled device.

For the sqs protocol, the endpoint is the ARN of an Amazon SQS queue.

For the application protocol, the endpoint is the EndpointArn of a mobile app and device.

For the lambda protocol, the endpoint is the ARN of an Lambda function.

For the firehose protocol, the endpoint is the ARN of an Amazon Kinesis Data Firehose delivery stream.

subscribe_topicArn :: Lens' Subscribe Text

The ARN of the topic you want to subscribe to.

subscribe_protocol :: Lens' Subscribe Text

The protocol that you want to use. Supported protocols include:

http – delivery of JSON-encoded message via HTTP POST

https – delivery of JSON-encoded message via HTTPS POST

email – delivery of message via SMTP

email-json – delivery of JSON-encoded message via SMTP

sms – delivery of message via SMS

sqs – delivery of JSON-encoded message to an Amazon SQS queue

application – delivery of JSON-encoded message to an EndpointArn for a mobile app and device

lambda – delivery of JSON-encoded message to an Lambda function

firehose – delivery of JSON-encoded message to an Amazon Kinesis Data Firehose delivery stream.

subscribeResponse_subscriptionArn :: Lens' SubscribeResponse (Maybe Text)

The ARN of the subscription if it is confirmed, or the string "pending confirmation" if the subscription requires confirmation. However, if the API request parameter ReturnSubscriptionArn is true, then the value is always the subscription ARN, even if the subscription requires confirmation.

subscribeResponse_httpStatus :: Lens' SubscribeResponse Int

The response's http status code.

ListOriginationNumbers

listOriginationNumbers_nextToken :: Lens' ListOriginationNumbers (Maybe Text)

Token that the previous ListOriginationNumbers request returns.

listOriginationNumbers_maxResults :: Lens' ListOriginationNumbers (Maybe Natural)

The maximum number of origination numbers to return.

listOriginationNumbersResponse_nextToken :: Lens' ListOriginationNumbersResponse (Maybe Text)

A NextToken string is returned when you call the ListOriginationNumbers operation if additional pages of records are available.

listOriginationNumbersResponse_phoneNumbers :: Lens' ListOriginationNumbersResponse (Maybe [PhoneNumberInformation])

A list of the calling account's verified and pending origination numbers.

listOriginationNumbersResponse_httpStatus :: Lens' ListOriginationNumbersResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the topic from which to remove tags.

untagResource_tagKeys :: Lens' UntagResource [Text]

The list of tag keys to remove from the specified topic.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

Unsubscribe

unsubscribe_subscriptionArn :: Lens' Unsubscribe Text

The ARN of the subscription to be deleted.

GetSMSSandboxAccountStatus

getSMSSandboxAccountStatusResponse_httpStatus :: Lens' GetSMSSandboxAccountStatusResponse Int

The response's http status code.

getSMSSandboxAccountStatusResponse_isInSandbox :: Lens' GetSMSSandboxAccountStatusResponse Bool

Indicates whether the calling account is in the SMS sandbox.

SetEndpointAttributes

setEndpointAttributes_endpointArn :: Lens' SetEndpointAttributes Text

EndpointArn used for SetEndpointAttributes action.

setEndpointAttributes_attributes :: Lens' SetEndpointAttributes (HashMap Text Text)

A map of the endpoint attributes. Attributes in this map include the following:

CustomUserData – arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

Enabled – flag that enables/disables delivery to the endpoint. Amazon SNS will set this to false when a notification service indicates to Amazon SNS that the endpoint is invalid. Users can set it back to true, typically after updating Token.

Token – device token, also referred to as a registration id, for an app and mobile device. This is returned from the notification service when an app and mobile device are registered with the notification service.

SetSubscriptionAttributes

setSubscriptionAttributes_attributeValue :: Lens' SetSubscriptionAttributes (Maybe Text)

The new value for the attribute in JSON format.

setSubscriptionAttributes_subscriptionArn :: Lens' SetSubscriptionAttributes Text

The ARN of the subscription to modify.

setSubscriptionAttributes_attributeName :: Lens' SetSubscriptionAttributes Text

A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that this action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

FilterPolicy – The simple JSON object that lets your subscriber receive only a subset of messages, rather than receiving every message published to the topic.

RawMessageDelivery – When set to true, enables raw message delivery to Amazon SQS or HTTP/S endpoints. This eliminates the need for the endpoints to process JSON formatting, which is otherwise created for Amazon SNS metadata.

RedrivePolicy – When specified, sends undeliverable messages to the specified Amazon SQS dead-letter queue. Messages that can't be delivered due to client errors (for example, when the subscribed endpoint is unreachable) or server errors (for example, when the service that powers the subscribed endpoint becomes unavailable) are held in the dead-letter queue for further analysis or reprocessing.

The following attribute applies only to Amazon Kinesis Data Firehose delivery stream subscriptions:

SubscriptionRoleArn – The ARN of the IAM role that has the following:

Permission to write to the Kinesis Data Firehose delivery stream

Amazon SNS listed as a trusted entity

Specifying a valid ARN for this attribute is required for Kinesis Data Firehose delivery stream subscriptions. For more information, see Fanout to Kinesis Data Firehose delivery streams in the Amazon SNS Developer Guide.

ConfirmSubscription

confirmSubscription_authenticateOnUnsubscribe :: Lens' ConfirmSubscription (Maybe Text)

Disallows unauthenticated unsubscribes of the subscription. If the value of this parameter is true and the request has an Amazon Web Services signature, then only the topic owner and the subscription owner can unsubscribe the endpoint. The unsubscribe action requires Amazon Web Services authentication.

confirmSubscription_topicArn :: Lens' ConfirmSubscription Text

The ARN of the topic for which you wish to confirm a subscription.

confirmSubscription_token :: Lens' ConfirmSubscription Text

Short-lived token sent to an endpoint during the Subscribe action.

confirmSubscriptionResponse_subscriptionArn :: Lens' ConfirmSubscriptionResponse (Maybe Text)

The ARN of the created subscription.

confirmSubscriptionResponse_httpStatus :: Lens' ConfirmSubscriptionResponse Int

The response's http status code.

Publish

publish_subject :: Lens' Publish (Maybe Text)

Optional parameter to be used as the "Subject" line when the message is delivered to email endpoints. This field will also be included, if present, in the standard JSON messages delivered to other endpoints.

Constraints: Subjects must be ASCII text that begins with a letter, number, or punctuation mark; must not include line breaks or control characters; and must be less than 100 characters long.

publish_targetArn :: Lens' Publish (Maybe Text)

If you don't specify a value for the TargetArn parameter, you must specify a value for the PhoneNumber or TopicArn parameters.

publish_messageAttributes :: Lens' Publish (Maybe (HashMap Text MessageAttributeValue))

Message attributes for Publish action.

publish_topicArn :: Lens' Publish (Maybe Text)

The topic you want to publish to.

If you don't specify a value for the TopicArn parameter, you must specify a value for the PhoneNumber or TargetArn parameters.

publish_phoneNumber :: Lens' Publish (Maybe Text)

The phone number to which you want to deliver an SMS message. Use E.164 format.

If you don't specify a value for the PhoneNumber parameter, you must specify a value for the TargetArn or TopicArn parameters.

publish_messageDeduplicationId :: Lens' Publish (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) topics. The MessageDeduplicationId can contain up to 128 alphanumeric characters (a-z, A-Z, 0-9) and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

Every message must have a unique MessageDeduplicationId, which is a token used for deduplication of sent messages. If a message with a particular MessageDeduplicationId is sent successfully, any message sent with the same MessageDeduplicationId during the 5-minute deduplication interval is treated as a duplicate.

If the topic has ContentBasedDeduplication set, the system generates a MessageDeduplicationId based on the contents of the message. Your MessageDeduplicationId overrides the generated one.

publish_messageStructure :: Lens' Publish (Maybe Text)

Set MessageStructure to json if you want to send a different message for each protocol. For example, using one publish action, you can send a short message to your SMS subscribers and a longer message to your email subscribers. If you set MessageStructure to json, the value of the Message parameter must:

be a syntactically valid JSON object; and

contain at least a top-level JSON key of "default" with a value that is a string.

You can define other top-level keys that define the message you want to send to a specific transport protocol (e.g., "http").

Valid value: json

publish_messageGroupId :: Lens' Publish (Maybe Text)

This parameter applies only to FIFO (first-in-first-out) topics. The MessageGroupId can contain up to 128 alphanumeric characters (a-z, A-Z, 0-9) and punctuation (!"#$%&'()*+,-./:;<=>?@[\]^_`{|}~).

The MessageGroupId is a tag that specifies that a message belongs to a specific message group. Messages that belong to the same message group are processed in a FIFO manner (however, messages in different message groups might be processed out of order). Every message must include a MessageGroupId.

publish_message :: Lens' Publish Text

The message you want to send.

If you are publishing to a topic and you want to send the same message to all transport protocols, include the text of the message as a String value. If you want to send different messages for each transport protocol, set the value of the MessageStructure parameter to json and use a JSON object for the Message parameter.

Constraints:

With the exception of SMS, messages must be UTF-8 encoded strings and at most 256 KB in size (262,144 bytes, not 262,144 characters).

For SMS, each message can contain up to 140 characters. This character limit depends on the encoding schema. For example, an SMS message can contain 160 GSM characters, 140 ASCII characters, or 70 UCS-2 characters.

If you publish a message that exceeds this size limit, Amazon SNS sends the message as multiple messages, each fitting within the size limit. Messages aren't truncated mid-word but are cut off at whole-word boundaries.

The total size limit for a single SMS Publish action is 1,600 characters.

JSON-specific constraints:

Keys in the JSON object that correspond to supported transport protocols must have simple JSON string values.

The values will be parsed (unescaped) before they are used in outgoing messages.

Outbound notifications are JSON encoded (meaning that the characters will be reescaped for sending).

Values have a minimum length of 0 (the empty string, "", is allowed).

Values have a maximum length bounded by the overall message size (so, including multiple protocols may limit message sizes).

Non-string values will cause the key to be ignored.

Keys that do not correspond to supported transport protocols are ignored.

Duplicate keys are not allowed.

Failure to parse or validate any key or value in the message will cause the Publish call to return an error (no partial delivery).

publishResponse_sequenceNumber :: Lens' PublishResponse (Maybe Text)

This response element applies only to FIFO (first-in-first-out) topics.

The sequence number is a large, non-consecutive number that Amazon SNS assigns to each message. The length of SequenceNumber is 128 bits. SequenceNumber continues to increase for each MessageGroupId.

publishResponse_messageId :: Lens' PublishResponse (Maybe Text)

Unique identifier assigned to the published message.

Length Constraint: Maximum 100 characters

publishResponse_httpStatus :: Lens' PublishResponse Int

The response's http status code.

Types

Endpoint

endpoint_attributes :: Lens' Endpoint (Maybe (HashMap Text Text))

Attributes for endpoint.

endpoint_endpointArn :: Lens' Endpoint (Maybe Text)

EndpointArn for mobile app and device.

MessageAttributeValue

messageAttributeValue_binaryValue :: Lens' MessageAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, for example, compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

messageAttributeValue_stringValue :: Lens' MessageAttributeValue (Maybe Text)

Strings are Unicode with UTF8 binary encoding. For a list of code values, see ASCII Printable Characters.

messageAttributeValue_dataType :: Lens' MessageAttributeValue Text

Amazon SNS supports the following logical data types: String, String.Array, Number, and Binary. For more information, see Message Attribute Data Types.

PhoneNumberInformation

phoneNumberInformation_status :: Lens' PhoneNumberInformation (Maybe Text)

The status of the phone number.

phoneNumberInformation_iso2CountryCode :: Lens' PhoneNumberInformation (Maybe Text)

The two-character code for the country or region, in ISO 3166-1 alpha-2 format.

phoneNumberInformation_createdAt :: Lens' PhoneNumberInformation (Maybe UTCTime)

The date and time when the phone number was created.

phoneNumberInformation_phoneNumber :: Lens' PhoneNumberInformation (Maybe Text)

The phone number.

phoneNumberInformation_numberCapabilities :: Lens' PhoneNumberInformation (Maybe [NumberCapability])

The capabilities of each phone number.

phoneNumberInformation_routeType :: Lens' PhoneNumberInformation (Maybe RouteType)

The list of supported routes.

PlatformApplication

platformApplication_platformApplicationArn :: Lens' PlatformApplication (Maybe Text)

PlatformApplicationArn for platform application object.

platformApplication_attributes :: Lens' PlatformApplication (Maybe (HashMap Text Text))

Attributes for platform application object.

SMSSandboxPhoneNumber

sMSSandboxPhoneNumber_status :: Lens' SMSSandboxPhoneNumber (Maybe SMSSandboxPhoneNumberVerificationStatus)

The destination phone number's verification status.

sMSSandboxPhoneNumber_phoneNumber :: Lens' SMSSandboxPhoneNumber (Maybe Text)

The destination phone number.

Subscription

subscription_protocol :: Lens' Subscription (Maybe Text)

The subscription's protocol.

subscription_owner :: Lens' Subscription (Maybe Text)

The subscription's owner.

subscription_topicArn :: Lens' Subscription (Maybe Text)

The ARN of the subscription's topic.

subscription_endpoint :: Lens' Subscription (Maybe Text)

The subscription's endpoint (format depends on the protocol).

subscription_subscriptionArn :: Lens' Subscription (Maybe Text)

The subscription's ARN.

Tag

tag_key :: Lens' Tag Text

The required key portion of the tag.

tag_value :: Lens' Tag Text

The optional value portion of the tag.

Topic

topic_topicArn :: Lens' Topic (Maybe Text)

The topic's ARN.