Amazonka.SNS.SetPlatformApplicationAttributes

Description

Sets the attributes of the platform application object for the supported push notification services, such as APNS and GCM (Firebase Cloud Messaging). For more information, see Using Amazon SNS Mobile Push Notifications. For information on configuring attributes for message delivery status, see Using Amazon SNS Application Attributes for Message Delivery Status.

Creating a Request

data SetPlatformApplicationAttributes

Input for SetPlatformApplicationAttributes action.

See: newSetPlatformApplicationAttributes smart constructor.

Constructors

SetPlatformApplicationAttributes'

Fields

platformApplicationArn :: Text

PlatformApplicationArn for SetPlatformApplicationAttributes action.

attributes :: HashMap Text Text

A map of the platform application attributes. Attributes in this map include the following:

PlatformCredential – The credential received from the notification service. For APNS and APNS_SANDBOX, PlatformCredential is private key. For GCM (Firebase Cloud Messaging), PlatformCredential is API key. For ADM, PlatformCredential is client secret.

PlatformPrincipal – The principal received from the notification service. For APNS and APNS_SANDBOX, PlatformPrincipal is SSL certificate. For GCM (Firebase Cloud Messaging), there is no PlatformPrincipal. For ADM, PlatformPrincipal is client id.

EventEndpointCreated – Topic ARN to which EndpointCreated event notifications are sent.

EventEndpointDeleted – Topic ARN to which EndpointDeleted event notifications are sent.

EventEndpointUpdated – Topic ARN to which EndpointUpdate event notifications are sent.

EventDeliveryFailure – Topic ARN to which DeliveryFailure event notifications are sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.

SuccessFeedbackRoleArn – IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf.

FailureFeedbackRoleArn – IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf.

SuccessFeedbackSampleRate – Sample rate percentage (0-100) of successfully delivered messages.

newSetPlatformApplicationAttributes

Arguments

:: Text
$sel:platformApplicationArn:SetPlatformApplicationAttributes'

-> SetPlatformApplicationAttributes

Create a value of SetPlatformApplicationAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platformApplicationArn:SetPlatformApplicationAttributes', setPlatformApplicationAttributes_platformApplicationArn - PlatformApplicationArn for SetPlatformApplicationAttributes action.

$sel:attributes:SetPlatformApplicationAttributes', setPlatformApplicationAttributes_attributes - A map of the platform application attributes. Attributes in this map include the following:

PlatformCredential – The credential received from the notification service. For APNS and APNS_SANDBOX, PlatformCredential is private key. For GCM (Firebase Cloud Messaging), PlatformCredential is API key. For ADM, PlatformCredential is client secret.

PlatformPrincipal – The principal received from the notification service. For APNS and APNS_SANDBOX, PlatformPrincipal is SSL certificate. For GCM (Firebase Cloud Messaging), there is no PlatformPrincipal. For ADM, PlatformPrincipal is client id.

EventEndpointCreated – Topic ARN to which EndpointCreated event notifications are sent.

EventEndpointDeleted – Topic ARN to which EndpointDeleted event notifications are sent.

EventEndpointUpdated – Topic ARN to which EndpointUpdate event notifications are sent.

EventDeliveryFailure – Topic ARN to which DeliveryFailure event notifications are sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.

SuccessFeedbackRoleArn – IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf.

FailureFeedbackRoleArn – IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf.

SuccessFeedbackSampleRate – Sample rate percentage (0-100) of successfully delivered messages.

Request Lenses

setPlatformApplicationAttributes_platformApplicationArn :: Lens' SetPlatformApplicationAttributes Text

PlatformApplicationArn for SetPlatformApplicationAttributes action.

setPlatformApplicationAttributes_attributes :: Lens' SetPlatformApplicationAttributes (HashMap Text Text)

A map of the platform application attributes. Attributes in this map include the following:

PlatformCredential – The credential received from the notification service. For APNS and APNS_SANDBOX, PlatformCredential is private key. For GCM (Firebase Cloud Messaging), PlatformCredential is API key. For ADM, PlatformCredential is client secret.

PlatformPrincipal – The principal received from the notification service. For APNS and APNS_SANDBOX, PlatformPrincipal is SSL certificate. For GCM (Firebase Cloud Messaging), there is no PlatformPrincipal. For ADM, PlatformPrincipal is client id.

EventEndpointCreated – Topic ARN to which EndpointCreated event notifications are sent.

EventEndpointDeleted – Topic ARN to which EndpointDeleted event notifications are sent.

EventEndpointUpdated – Topic ARN to which EndpointUpdate event notifications are sent.

EventDeliveryFailure – Topic ARN to which DeliveryFailure event notifications are sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.

SuccessFeedbackRoleArn – IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf.

FailureFeedbackRoleArn – IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf.

SuccessFeedbackSampleRate – Sample rate percentage (0-100) of successfully delivered messages.

Destructuring the Response

data SetPlatformApplicationAttributesResponse

See: newSetPlatformApplicationAttributesResponse smart constructor.

Constructors

SetPlatformApplicationAttributesResponse'

newSetPlatformApplicationAttributesResponse :: SetPlatformApplicationAttributesResponse

Create a value of SetPlatformApplicationAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.