Amazonka.SNS.SetSubscriptionAttributes

Description

Allows a subscription owner to set an attribute of the subscription to a new value.

Creating a Request

data SetSubscriptionAttributes

Input for SetSubscriptionAttributes action.

See: newSetSubscriptionAttributes smart constructor.

Constructors

SetSubscriptionAttributes'

Fields

attributeValue :: Maybe Text

The new value for the attribute in JSON format.

subscriptionArn :: Text

The ARN of the subscription to modify.

attributeName :: Text

A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that this action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

FilterPolicy – The simple JSON object that lets your subscriber receive only a subset of messages, rather than receiving every message published to the topic.

RawMessageDelivery – When set to true, enables raw message delivery to Amazon SQS or HTTP/S endpoints. This eliminates the need for the endpoints to process JSON formatting, which is otherwise created for Amazon SNS metadata.

RedrivePolicy – When specified, sends undeliverable messages to the specified Amazon SQS dead-letter queue. Messages that can't be delivered due to client errors (for example, when the subscribed endpoint is unreachable) or server errors (for example, when the service that powers the subscribed endpoint becomes unavailable) are held in the dead-letter queue for further analysis or reprocessing.

The following attribute applies only to Amazon Kinesis Data Firehose delivery stream subscriptions:

SubscriptionRoleArn – The ARN of the IAM role that has the following:

Permission to write to the Kinesis Data Firehose delivery stream

Amazon SNS listed as a trusted entity

Specifying a valid ARN for this attribute is required for Kinesis Data Firehose delivery stream subscriptions. For more information, see Fanout to Kinesis Data Firehose delivery streams in the Amazon SNS Developer Guide.

newSetSubscriptionAttributes

Arguments

:: Text
$sel:subscriptionArn:SetSubscriptionAttributes'

-> Text
$sel:attributeName:SetSubscriptionAttributes'

-> SetSubscriptionAttributes

Create a value of SetSubscriptionAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeValue:SetSubscriptionAttributes', setSubscriptionAttributes_attributeValue - The new value for the attribute in JSON format.

$sel:subscriptionArn:SetSubscriptionAttributes', setSubscriptionAttributes_subscriptionArn - The ARN of the subscription to modify.

$sel:attributeName:SetSubscriptionAttributes', setSubscriptionAttributes_attributeName - A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that this action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

FilterPolicy – The simple JSON object that lets your subscriber receive only a subset of messages, rather than receiving every message published to the topic.

RawMessageDelivery – When set to true, enables raw message delivery to Amazon SQS or HTTP/S endpoints. This eliminates the need for the endpoints to process JSON formatting, which is otherwise created for Amazon SNS metadata.

RedrivePolicy – When specified, sends undeliverable messages to the specified Amazon SQS dead-letter queue. Messages that can't be delivered due to client errors (for example, when the subscribed endpoint is unreachable) or server errors (for example, when the service that powers the subscribed endpoint becomes unavailable) are held in the dead-letter queue for further analysis or reprocessing.

The following attribute applies only to Amazon Kinesis Data Firehose delivery stream subscriptions:

SubscriptionRoleArn – The ARN of the IAM role that has the following:

Permission to write to the Kinesis Data Firehose delivery stream

Amazon SNS listed as a trusted entity

Specifying a valid ARN for this attribute is required for Kinesis Data Firehose delivery stream subscriptions. For more information, see Fanout to Kinesis Data Firehose delivery streams in the Amazon SNS Developer Guide.

Request Lenses

setSubscriptionAttributes_attributeValue :: Lens' SetSubscriptionAttributes (Maybe Text)

The new value for the attribute in JSON format.

setSubscriptionAttributes_subscriptionArn :: Lens' SetSubscriptionAttributes Text

The ARN of the subscription to modify.

setSubscriptionAttributes_attributeName :: Lens' SetSubscriptionAttributes Text

A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that this action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

FilterPolicy – The simple JSON object that lets your subscriber receive only a subset of messages, rather than receiving every message published to the topic.

RawMessageDelivery – When set to true, enables raw message delivery to Amazon SQS or HTTP/S endpoints. This eliminates the need for the endpoints to process JSON formatting, which is otherwise created for Amazon SNS metadata.

RedrivePolicy – When specified, sends undeliverable messages to the specified Amazon SQS dead-letter queue. Messages that can't be delivered due to client errors (for example, when the subscribed endpoint is unreachable) or server errors (for example, when the service that powers the subscribed endpoint becomes unavailable) are held in the dead-letter queue for further analysis or reprocessing.

The following attribute applies only to Amazon Kinesis Data Firehose delivery stream subscriptions:

SubscriptionRoleArn – The ARN of the IAM role that has the following:

Permission to write to the Kinesis Data Firehose delivery stream

Amazon SNS listed as a trusted entity

Specifying a valid ARN for this attribute is required for Kinesis Data Firehose delivery stream subscriptions. For more information, see Fanout to Kinesis Data Firehose delivery streams in the Amazon SNS Developer Guide.

Destructuring the Response

data SetSubscriptionAttributesResponse

See: newSetSubscriptionAttributesResponse smart constructor.

Constructors

SetSubscriptionAttributesResponse'

newSetSubscriptionAttributesResponse :: SetSubscriptionAttributesResponse

Create a value of SetSubscriptionAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.