Amazonka.SNS.Subscribe

Description

Subscribes an endpoint to an Amazon SNS topic. If the endpoint type is HTTP/S or email, or if the endpoint and the topic are not in the same account, the endpoint owner must run the ConfirmSubscription action to confirm the subscription.

You call the ConfirmSubscription action with the token from the subscription response. Confirmation tokens are valid for three days.

This action is throttled at 100 transactions per second (TPS).

Creating a Request

data Subscribe

Input for Subscribe action.

See: newSubscribe smart constructor.

Constructors

Subscribe'

Fields

returnSubscriptionArn :: Maybe Bool

Sets whether the response from the Subscribe request includes the subscription ARN, even if the subscription is not yet confirmed.

If you set this parameter to true, the response includes the ARN in all cases, even if the subscription is not yet confirmed. In addition to the ARN for confirmed subscriptions, the response also includes the pending subscription ARN value for subscriptions that aren't yet confirmed. A subscription becomes confirmed when the subscriber calls the ConfirmSubscription action with a confirmation token.

The default value is false.

attributes :: Maybe (HashMap Text Text)

A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that the Subscribe action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

FilterPolicy – The simple JSON object that lets your subscriber receive only a subset of messages, rather than receiving every message published to the topic.

RawMessageDelivery – When set to true, enables raw message delivery to Amazon SQS or HTTP/S endpoints. This eliminates the need for the endpoints to process JSON formatting, which is otherwise created for Amazon SNS metadata.

RedrivePolicy – When specified, sends undeliverable messages to the specified Amazon SQS dead-letter queue. Messages that can't be delivered due to client errors (for example, when the subscribed endpoint is unreachable) or server errors (for example, when the service that powers the subscribed endpoint becomes unavailable) are held in the dead-letter queue for further analysis or reprocessing.

The following attribute applies only to Amazon Kinesis Data Firehose delivery stream subscriptions:

SubscriptionRoleArn – The ARN of the IAM role that has the following:

Permission to write to the Kinesis Data Firehose delivery stream

Amazon SNS listed as a trusted entity

Specifying a valid ARN for this attribute is required for Kinesis Data Firehose delivery stream subscriptions. For more information, see Fanout to Kinesis Data Firehose delivery streams in the Amazon SNS Developer Guide.

endpoint :: Maybe Text

The endpoint that you want to receive notifications. Endpoints vary by protocol:

For the http protocol, the (public) endpoint is a URL beginning with http://.

For the https protocol, the (public) endpoint is a URL beginning with https://.

For the email protocol, the endpoint is an email address.

For the email-json protocol, the endpoint is an email address.

For the sms protocol, the endpoint is a phone number of an SMS-enabled device.

For the sqs protocol, the endpoint is the ARN of an Amazon SQS queue.

For the application protocol, the endpoint is the EndpointArn of a mobile app and device.

For the lambda protocol, the endpoint is the ARN of an Lambda function.

For the firehose protocol, the endpoint is the ARN of an Amazon Kinesis Data Firehose delivery stream.

topicArn :: Text

The ARN of the topic you want to subscribe to.

protocol :: Text

The protocol that you want to use. Supported protocols include:

http – delivery of JSON-encoded message via HTTP POST

https – delivery of JSON-encoded message via HTTPS POST

email – delivery of message via SMTP

email-json – delivery of JSON-encoded message via SMTP

sms – delivery of message via SMS

sqs – delivery of JSON-encoded message to an Amazon SQS queue

application – delivery of JSON-encoded message to an EndpointArn for a mobile app and device

lambda – delivery of JSON-encoded message to an Lambda function

firehose – delivery of JSON-encoded message to an Amazon Kinesis Data Firehose delivery stream.

newSubscribe

Arguments

:: Text
$sel:topicArn:Subscribe'

-> Text
$sel:protocol:Subscribe'

-> Subscribe

Create a value of Subscribe with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:returnSubscriptionArn:Subscribe', subscribe_returnSubscriptionArn - Sets whether the response from the Subscribe request includes the subscription ARN, even if the subscription is not yet confirmed.

If you set this parameter to true, the response includes the ARN in all cases, even if the subscription is not yet confirmed. In addition to the ARN for confirmed subscriptions, the response also includes the pending subscription ARN value for subscriptions that aren't yet confirmed. A subscription becomes confirmed when the subscriber calls the ConfirmSubscription action with a confirmation token.

The default value is false.

$sel:attributes:Subscribe', subscribe_attributes - A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that the Subscribe action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

FilterPolicy – The simple JSON object that lets your subscriber receive only a subset of messages, rather than receiving every message published to the topic.

RawMessageDelivery – When set to true, enables raw message delivery to Amazon SQS or HTTP/S endpoints. This eliminates the need for the endpoints to process JSON formatting, which is otherwise created for Amazon SNS metadata.

RedrivePolicy – When specified, sends undeliverable messages to the specified Amazon SQS dead-letter queue. Messages that can't be delivered due to client errors (for example, when the subscribed endpoint is unreachable) or server errors (for example, when the service that powers the subscribed endpoint becomes unavailable) are held in the dead-letter queue for further analysis or reprocessing.

The following attribute applies only to Amazon Kinesis Data Firehose delivery stream subscriptions:

SubscriptionRoleArn – The ARN of the IAM role that has the following:

Permission to write to the Kinesis Data Firehose delivery stream

Amazon SNS listed as a trusted entity

Specifying a valid ARN for this attribute is required for Kinesis Data Firehose delivery stream subscriptions. For more information, see Fanout to Kinesis Data Firehose delivery streams in the Amazon SNS Developer Guide.

$sel:endpoint:Subscribe', subscribe_endpoint - The endpoint that you want to receive notifications. Endpoints vary by protocol:

For the http protocol, the (public) endpoint is a URL beginning with http://.

For the https protocol, the (public) endpoint is a URL beginning with https://.

For the email protocol, the endpoint is an email address.

For the email-json protocol, the endpoint is an email address.

For the sms protocol, the endpoint is a phone number of an SMS-enabled device.

For the sqs protocol, the endpoint is the ARN of an Amazon SQS queue.

For the application protocol, the endpoint is the EndpointArn of a mobile app and device.

For the lambda protocol, the endpoint is the ARN of an Lambda function.

For the firehose protocol, the endpoint is the ARN of an Amazon Kinesis Data Firehose delivery stream.

$sel:topicArn:Subscribe', subscribe_topicArn - The ARN of the topic you want to subscribe to.

$sel:protocol:Subscribe', subscribe_protocol - The protocol that you want to use. Supported protocols include:

http – delivery of JSON-encoded message via HTTP POST

https – delivery of JSON-encoded message via HTTPS POST

email – delivery of message via SMTP

email-json – delivery of JSON-encoded message via SMTP

sms – delivery of message via SMS

sqs – delivery of JSON-encoded message to an Amazon SQS queue

application – delivery of JSON-encoded message to an EndpointArn for a mobile app and device

lambda – delivery of JSON-encoded message to an Lambda function

firehose – delivery of JSON-encoded message to an Amazon Kinesis Data Firehose delivery stream.

Request Lenses

subscribe_returnSubscriptionArn :: Lens' Subscribe (Maybe Bool)

Sets whether the response from the Subscribe request includes the subscription ARN, even if the subscription is not yet confirmed.

If you set this parameter to true, the response includes the ARN in all cases, even if the subscription is not yet confirmed. In addition to the ARN for confirmed subscriptions, the response also includes the pending subscription ARN value for subscriptions that aren't yet confirmed. A subscription becomes confirmed when the subscriber calls the ConfirmSubscription action with a confirmation token.

The default value is false.

subscribe_attributes :: Lens' Subscribe (Maybe (HashMap Text Text))

A map of attributes with their corresponding values.

The following lists the names, descriptions, and values of the special request parameters that the Subscribe action uses:

DeliveryPolicy – The policy that defines how Amazon SNS retries failed deliveries to HTTP/S endpoints.

FilterPolicy – The simple JSON object that lets your subscriber receive only a subset of messages, rather than receiving every message published to the topic.

RawMessageDelivery – When set to true, enables raw message delivery to Amazon SQS or HTTP/S endpoints. This eliminates the need for the endpoints to process JSON formatting, which is otherwise created for Amazon SNS metadata.

RedrivePolicy – When specified, sends undeliverable messages to the specified Amazon SQS dead-letter queue. Messages that can't be delivered due to client errors (for example, when the subscribed endpoint is unreachable) or server errors (for example, when the service that powers the subscribed endpoint becomes unavailable) are held in the dead-letter queue for further analysis or reprocessing.

The following attribute applies only to Amazon Kinesis Data Firehose delivery stream subscriptions:

SubscriptionRoleArn – The ARN of the IAM role that has the following:

Permission to write to the Kinesis Data Firehose delivery stream

Amazon SNS listed as a trusted entity

Specifying a valid ARN for this attribute is required for Kinesis Data Firehose delivery stream subscriptions. For more information, see Fanout to Kinesis Data Firehose delivery streams in the Amazon SNS Developer Guide.

subscribe_endpoint :: Lens' Subscribe (Maybe Text)

The endpoint that you want to receive notifications. Endpoints vary by protocol:

For the http protocol, the (public) endpoint is a URL beginning with http://.

For the https protocol, the (public) endpoint is a URL beginning with https://.

For the email protocol, the endpoint is an email address.

For the email-json protocol, the endpoint is an email address.

For the sms protocol, the endpoint is a phone number of an SMS-enabled device.

For the sqs protocol, the endpoint is the ARN of an Amazon SQS queue.

For the application protocol, the endpoint is the EndpointArn of a mobile app and device.

For the lambda protocol, the endpoint is the ARN of an Lambda function.

For the firehose protocol, the endpoint is the ARN of an Amazon Kinesis Data Firehose delivery stream.

subscribe_topicArn :: Lens' Subscribe Text

The ARN of the topic you want to subscribe to.

subscribe_protocol :: Lens' Subscribe Text

The protocol that you want to use. Supported protocols include:

http – delivery of JSON-encoded message via HTTP POST

https – delivery of JSON-encoded message via HTTPS POST

email – delivery of message via SMTP

email-json – delivery of JSON-encoded message via SMTP

sms – delivery of message via SMS

sqs – delivery of JSON-encoded message to an Amazon SQS queue

application – delivery of JSON-encoded message to an EndpointArn for a mobile app and device

lambda – delivery of JSON-encoded message to an Lambda function

firehose – delivery of JSON-encoded message to an Amazon Kinesis Data Firehose delivery stream.

Destructuring the Response

data SubscribeResponse

Response for Subscribe action.

See: newSubscribeResponse smart constructor.

Constructors

SubscribeResponse'

Fields

subscriptionArn :: Maybe Text

The ARN of the subscription if it is confirmed, or the string "pending confirmation" if the subscription requires confirmation. However, if the API request parameter ReturnSubscriptionArn is true, then the value is always the subscription ARN, even if the subscription requires confirmation.

httpStatus :: Int

The response's http status code.

newSubscribeResponse

Arguments

:: Int
$sel:httpStatus:SubscribeResponse'

-> SubscribeResponse

Create a value of SubscribeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionArn:SubscribeResponse', subscribeResponse_subscriptionArn - The ARN of the subscription if it is confirmed, or the string "pending confirmation" if the subscription requires confirmation. However, if the API request parameter ReturnSubscriptionArn is true, then the value is always the subscription ARN, even if the subscription requires confirmation.

$sel:httpStatus:SubscribeResponse', subscribeResponse_httpStatus - The response's http status code.

Response Lenses

subscribeResponse_subscriptionArn :: Lens' SubscribeResponse (Maybe Text)

The ARN of the subscription if it is confirmed, or the string "pending confirmation" if the subscription requires confirmation. However, if the API request parameter ReturnSubscriptionArn is true, then the value is always the subscription ARN, even if the subscription requires confirmation.

subscribeResponse_httpStatus :: Lens' SubscribeResponse Int

The response's http status code.