Amazonka.SNS.VerifySMSSandboxPhoneNumber

Description

Verifies a destination phone number with a one-time password (OTP) for the calling account.

When you start using Amazon SNS to send SMS messages, your account is in the SMS sandbox. The SMS sandbox provides a safe environment for you to try Amazon SNS features without risking your reputation as an SMS sender. While your account is in the SMS sandbox, you can use all of the features of Amazon SNS. However, you can send SMS messages only to verified destination phone numbers. For more information, including how to move out of the sandbox to send messages without restrictions, see SMS sandbox in the Amazon SNS Developer Guide.

Creating a Request

data VerifySMSSandboxPhoneNumber

See: newVerifySMSSandboxPhoneNumber smart constructor.

Constructors

VerifySMSSandboxPhoneNumber'

Fields

phoneNumber :: Text

The destination phone number to verify.

oneTimePassword :: Text

The OTP sent to the destination number from the CreateSMSSandBoxPhoneNumber call.

newVerifySMSSandboxPhoneNumber

Arguments

:: Text
$sel:phoneNumber:VerifySMSSandboxPhoneNumber'

-> Text
$sel:oneTimePassword:VerifySMSSandboxPhoneNumber'

-> VerifySMSSandboxPhoneNumber

Create a value of VerifySMSSandboxPhoneNumber with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumber:VerifySMSSandboxPhoneNumber', verifySMSSandboxPhoneNumber_phoneNumber - The destination phone number to verify.

$sel:oneTimePassword:VerifySMSSandboxPhoneNumber', verifySMSSandboxPhoneNumber_oneTimePassword - The OTP sent to the destination number from the CreateSMSSandBoxPhoneNumber call.

Request Lenses

verifySMSSandboxPhoneNumber_phoneNumber :: Lens' VerifySMSSandboxPhoneNumber Text

The destination phone number to verify.

verifySMSSandboxPhoneNumber_oneTimePassword :: Lens' VerifySMSSandboxPhoneNumber Text

The OTP sent to the destination number from the CreateSMSSandBoxPhoneNumber call.

Destructuring the Response

data VerifySMSSandboxPhoneNumberResponse

The destination phone number's verification status.

See: newVerifySMSSandboxPhoneNumberResponse smart constructor.

Constructors

VerifySMSSandboxPhoneNumberResponse'

Fields

httpStatus :: Int

The response's http status code.

newVerifySMSSandboxPhoneNumberResponse

Arguments

:: Int
$sel:httpStatus:VerifySMSSandboxPhoneNumberResponse'

-> VerifySMSSandboxPhoneNumberResponse

Create a value of VerifySMSSandboxPhoneNumberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:VerifySMSSandboxPhoneNumberResponse', verifySMSSandboxPhoneNumberResponse_httpStatus - The response's http status code.

Response Lenses

verifySMSSandboxPhoneNumberResponse_httpStatus :: Lens' VerifySMSSandboxPhoneNumberResponse Int

The response's http status code.