Amazonka.SQS.GetQueueUrl

Description

Returns the URL of an existing Amazon SQS queue.

To access a queue that belongs to another AWS account, use the QueueOwnerAWSAccountId parameter to specify the account ID of the queue's owner. The queue's owner must grant you permission to access the queue. For more information about shared queue access, see AddPermission or see Allow Developers to Write Messages to a Shared Queue in the Amazon SQS Developer Guide.

Creating a Request

data GetQueueUrl

See: newGetQueueUrl smart constructor.

Constructors

GetQueueUrl'

Fields

queueOwnerAWSAccountId :: Maybe Text

The account ID of the account that created the queue.

queueName :: Text

The name of the queue whose URL must be fetched. Maximum 80 characters. Valid values: alphanumeric characters, hyphens (-), and underscores (_).

Queue URLs and names are case-sensitive.

newGetQueueUrl

Arguments

:: Text
$sel:queueName:GetQueueUrl'

-> GetQueueUrl

Create a value of GetQueueUrl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueOwnerAWSAccountId:GetQueueUrl', getQueueUrl_queueOwnerAWSAccountId - The account ID of the account that created the queue.

$sel:queueName:GetQueueUrl', getQueueUrl_queueName - The name of the queue whose URL must be fetched. Maximum 80 characters. Valid values: alphanumeric characters, hyphens (-), and underscores (_).

Queue URLs and names are case-sensitive.

Request Lenses

getQueueUrl_queueOwnerAWSAccountId :: Lens' GetQueueUrl (Maybe Text)

The account ID of the account that created the queue.

getQueueUrl_queueName :: Lens' GetQueueUrl Text

The name of the queue whose URL must be fetched. Maximum 80 characters. Valid values: alphanumeric characters, hyphens (-), and underscores (_).

Queue URLs and names are case-sensitive.

Destructuring the Response

data GetQueueUrlResponse

For more information, see Interpreting Responses in the Amazon SQS Developer Guide.

See: newGetQueueUrlResponse smart constructor.

Constructors

GetQueueUrlResponse'

Fields

httpStatus :: Int

The response's http status code.

queueUrl :: Text

The URL of the queue.

newGetQueueUrlResponse

Arguments

:: Int
$sel:httpStatus:GetQueueUrlResponse'

-> Text
$sel:queueUrl:GetQueueUrlResponse'

-> GetQueueUrlResponse

Create a value of GetQueueUrlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetQueueUrlResponse', getQueueUrlResponse_httpStatus - The response's http status code.

$sel:queueUrl:GetQueueUrlResponse', getQueueUrlResponse_queueUrl - The URL of the queue.

Response Lenses

getQueueUrlResponse_httpStatus :: Lens' GetQueueUrlResponse Int

The response's http status code.

getQueueUrlResponse_queueUrl :: Lens' GetQueueUrlResponse Text

The URL of the queue.