Amazonka.SQS.ListQueues

Description

Returns a list of your queues in the current region. The response includes a maximum of 1,000 results. If you specify a value for the optional QueueNamePrefix parameter, only queues with a name that begins with the specified value are returned.

The listQueues methods supports pagination. Set parameter MaxResults in the request to specify the maximum number of results to be returned in the response. If you do not set MaxResults, the response includes a maximum of 1,000 results. If you set MaxResults and there are additional results to display, the response includes a value for NextToken. Use NextToken as a parameter in your next request to listQueues to receive the next page of results.

Cross-account permissions don't apply to this action. For more information, see Grant cross-account permissions to a role and a user name in the Amazon SQS Developer Guide.

This operation returns paginated results.

Creating a Request

data ListQueues

See: newListQueues smart constructor.

Constructors

ListQueues'

Fields

queueNamePrefix :: Maybe Text

A string to use for filtering the list results. Only those queues whose name begins with the specified string are returned.

Queue URLs and names are case-sensitive.

nextToken :: Maybe Text

Pagination token to request the next set of results.

maxResults :: Maybe Int

Maximum number of results to include in the response. Value range is 1 to 1000. You must set MaxResults to receive a value for NextToken in the response.

newListQueues :: ListQueues

Create a value of ListQueues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueNamePrefix:ListQueues', listQueues_queueNamePrefix - A string to use for filtering the list results. Only those queues whose name begins with the specified string are returned.

Queue URLs and names are case-sensitive.

$sel:nextToken:ListQueues', listQueues_nextToken - Pagination token to request the next set of results.

$sel:maxResults:ListQueues', listQueues_maxResults - Maximum number of results to include in the response. Value range is 1 to 1000. You must set MaxResults to receive a value for NextToken in the response.

Request Lenses

listQueues_queueNamePrefix :: Lens' ListQueues (Maybe Text)

A string to use for filtering the list results. Only those queues whose name begins with the specified string are returned.

Queue URLs and names are case-sensitive.

listQueues_nextToken :: Lens' ListQueues (Maybe Text)

Pagination token to request the next set of results.

listQueues_maxResults :: Lens' ListQueues (Maybe Int)

Maximum number of results to include in the response. Value range is 1 to 1000. You must set MaxResults to receive a value for NextToken in the response.

Destructuring the Response

data ListQueuesResponse

A list of your queues.

See: newListQueuesResponse smart constructor.

Constructors

ListQueuesResponse'

Fields

queueUrls :: Maybe [Text]

A list of queue URLs, up to 1,000 entries, or the value of MaxResults that you sent in the request.

nextToken :: Maybe Text

Pagination token to include in the next request. Token value is null if there are no additional results to request, or if you did not set MaxResults in the request.

httpStatus :: Int

The response's http status code.

newListQueuesResponse

Arguments

:: Int
$sel:httpStatus:ListQueuesResponse'

-> ListQueuesResponse

Create a value of ListQueuesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queueUrls:ListQueuesResponse', listQueuesResponse_queueUrls - A list of queue URLs, up to 1,000 entries, or the value of MaxResults that you sent in the request.

$sel:nextToken:ListQueues', listQueuesResponse_nextToken - Pagination token to include in the next request. Token value is null if there are no additional results to request, or if you did not set MaxResults in the request.

$sel:httpStatus:ListQueuesResponse', listQueuesResponse_httpStatus - The response's http status code.

Response Lenses

listQueuesResponse_queueUrls :: Lens' ListQueuesResponse (Maybe [Text])

A list of queue URLs, up to 1,000 entries, or the value of MaxResults that you sent in the request.

listQueuesResponse_nextToken :: Lens' ListQueuesResponse (Maybe Text)

Pagination token to include in the next request. Token value is null if there are no additional results to request, or if you did not set MaxResults in the request.

listQueuesResponse_httpStatus :: Lens' ListQueuesResponse Int

The response's http status code.