Amazonka.SQS.Types.Message

Description

Documentation

data Message

An Amazon SQS message.

See: newMessage smart constructor.

Constructors

Message'

Fields

messageAttributes :: Maybe (HashMap Text MessageAttributeValue)

Each message attribute consists of a Name, Type, and Value. For more information, see Amazon SQS message attributes in the Amazon SQS Developer Guide.

mD5OfBody :: Maybe Text

An MD5 digest of the non-URL-encoded message body string.

body :: Maybe Text

The message's contents (not URL-encoded).

attributes :: Maybe (HashMap MessageAttribute Text)

A map of the attributes requested in ReceiveMessage to their respective values. Supported attributes:

ApproximateReceiveCount

ApproximateFirstReceiveTimestamp

MessageDeduplicationId

MessageGroupId

SenderId

SentTimestamp

SequenceNumber

ApproximateFirstReceiveTimestamp and SentTimestamp are each returned as an integer representing the epoch time in milliseconds.

receiptHandle :: Maybe Text

An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

messageId :: Maybe Text

A unique identifier for the message. A MessageIdis considered unique across all accounts for an extended period of time.

mD5OfMessageAttributes :: Maybe Text

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

newMessage :: Message

Create a value of Message with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageAttributes:Message', message_messageAttributes - Each message attribute consists of a Name, Type, and Value. For more information, see Amazon SQS message attributes in the Amazon SQS Developer Guide.

$sel:mD5OfBody:Message', message_mD5OfBody - An MD5 digest of the non-URL-encoded message body string.

$sel:body:Message', message_body - The message's contents (not URL-encoded).

$sel:attributes:Message', message_attributes - A map of the attributes requested in ReceiveMessage to their respective values. Supported attributes:

ApproximateReceiveCount

ApproximateFirstReceiveTimestamp

MessageDeduplicationId

MessageGroupId

SenderId

SentTimestamp

SequenceNumber

ApproximateFirstReceiveTimestamp and SentTimestamp are each returned as an integer representing the epoch time in milliseconds.

$sel:receiptHandle:Message', message_receiptHandle - An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

$sel:messageId:Message', message_messageId - A unique identifier for the message. A MessageIdis considered unique across all accounts for an extended period of time.

$sel:mD5OfMessageAttributes:Message', message_mD5OfMessageAttributes - An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.

message_messageAttributes :: Lens' Message (Maybe (HashMap Text MessageAttributeValue))

Each message attribute consists of a Name, Type, and Value. For more information, see Amazon SQS message attributes in the Amazon SQS Developer Guide.

message_mD5OfBody :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message body string.

message_body :: Lens' Message (Maybe Text)

The message's contents (not URL-encoded).

message_attributes :: Lens' Message (Maybe (HashMap MessageAttribute Text))

A map of the attributes requested in ReceiveMessage to their respective values. Supported attributes:

ApproximateReceiveCount

ApproximateFirstReceiveTimestamp

MessageDeduplicationId

MessageGroupId

SenderId

SentTimestamp

SequenceNumber

ApproximateFirstReceiveTimestamp and SentTimestamp are each returned as an integer representing the epoch time in milliseconds.

message_receiptHandle :: Lens' Message (Maybe Text)

An identifier associated with the act of receiving the message. A new receipt handle is returned every time you receive a message. When deleting a message, you provide the last received receipt handle to delete the message.

message_messageId :: Lens' Message (Maybe Text)

A unique identifier for the message. A MessageIdis considered unique across all accounts for an extended period of time.

message_mD5OfMessageAttributes :: Lens' Message (Maybe Text)

An MD5 digest of the non-URL-encoded message attribute string. You can use this attribute to verify that Amazon SQS received the message correctly. Amazon SQS URL-decodes the message before creating the MD5 digest. For information about MD5, see RFC1321.