Amazonka.SQS.Types.MessageSystemAttributeValue

Description

Documentation

data MessageSystemAttributeValue

The user-specified message system attribute value. For string data types, the Value attribute has the same restrictions on the content as the message body. For more information, see SendMessage.

Name, type, value and the message body must not be empty or null.

See: newMessageSystemAttributeValue smart constructor.

Constructors

MessageSystemAttributeValue'

Fields

binaryValue :: Maybe Base64

Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.

stringListValues :: Maybe [Text]

Not implemented. Reserved for future use.

stringValue :: Maybe Text

Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters.

binaryListValues :: Maybe [Base64]

Not implemented. Reserved for future use.

dataType :: Text

Amazon SQS supports the following logical data types: String, Number, and Binary. For the Number data type, you must use StringValue.

You can also append custom labels. For more information, see Amazon SQS Message Attributes in the Amazon SQS Developer Guide.

newMessageSystemAttributeValue

Arguments

:: Text
$sel:dataType:MessageSystemAttributeValue'

-> MessageSystemAttributeValue

Create a value of MessageSystemAttributeValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:binaryValue:MessageSystemAttributeValue', messageSystemAttributeValue_binaryValue - Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:stringListValues:MessageSystemAttributeValue', messageSystemAttributeValue_stringListValues - Not implemented. Reserved for future use.

$sel:stringValue:MessageSystemAttributeValue', messageSystemAttributeValue_stringValue - Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters.

$sel:binaryListValues:MessageSystemAttributeValue', messageSystemAttributeValue_binaryListValues - Not implemented. Reserved for future use.

$sel:dataType:MessageSystemAttributeValue', messageSystemAttributeValue_dataType - Amazon SQS supports the following logical data types: String, Number, and Binary. For the Number data type, you must use StringValue.

You can also append custom labels. For more information, see Amazon SQS Message Attributes in the Amazon SQS Developer Guide.

messageSystemAttributeValue_binaryValue :: Lens' MessageSystemAttributeValue (Maybe ByteString)

Binary type attributes can store any binary data, such as compressed data, encrypted data, or images.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

messageSystemAttributeValue_stringListValues :: Lens' MessageSystemAttributeValue (Maybe [Text])

Not implemented. Reserved for future use.

messageSystemAttributeValue_stringValue :: Lens' MessageSystemAttributeValue (Maybe Text)

Strings are Unicode with UTF-8 binary encoding. For a list of code values, see ASCII Printable Characters.

messageSystemAttributeValue_binaryListValues :: Lens' MessageSystemAttributeValue (Maybe [ByteString])

Not implemented. Reserved for future use.

messageSystemAttributeValue_dataType :: Lens' MessageSystemAttributeValue Text

Amazon SQS supports the following logical data types: String, Number, and Binary. For the Number data type, you must use StringValue.

You can also append custom labels. For more information, see Amazon SQS Message Attributes in the Amazon SQS Developer Guide.