Amazonka.SSM.CreateDocument

Description

Creates a Amazon Web Services Systems Manager (SSM document). An SSM document defines the actions that Systems Manager performs on your managed instances. For more information about SSM documents, including information about supported schemas, features, and syntax, see Amazon Web Services Systems Manager Documents in the Amazon Web Services Systems Manager User Guide.

Creating a Request

data CreateDocument

See: newCreateDocument smart constructor.

Constructors

CreateDocument'

Fields

documentType :: Maybe DocumentType

The type of document to create.

attachments :: Maybe [AttachmentsSource]

A list of key-value pairs that describe attachments to a version of a document.

versionName :: Maybe Text

An optional field specifying the version of the artifact you are creating with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

targetType :: Maybe Text

Specify a target type to define the kinds of resources the document can run on. For example, to run a document on EC2 instances, specify the following value: /AWS::EC2::Instance. If you specify a value of '/' the document can run on all types of resources. If you don't specify a value, the document can't run on any resources. For a list of valid resource types, see Amazon Web Services resource and property types reference in the CloudFormation User Guide.

documentFormat :: Maybe DocumentFormat

Specify the document format for the request. The document format can be JSON, YAML, or TEXT. JSON is the default format.

displayName :: Maybe Text

An optional field where you can specify a friendly name for the SSM document. This value can differ for each version of the document. You can update this value at a later time using the UpdateDocument operation.

requires :: Maybe (NonEmpty DocumentRequires)

A list of SSM documents required by a document. This parameter is used exclusively by AppConfig. When a user creates an AppConfig configuration in an SSM document, the user must also specify a required document for validation purposes. In this case, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document for validation purposes. For more information, see What is AppConfig? in the AppConfig User Guide.

tags :: Maybe [Tag]

Optional metadata that you assign to a resource. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag an SSM document to identify the types of targets or the environment where it will run. In this case, you could specify the following key-value pairs:

Key=OS,Value=Windows

Key=Environment,Value=Production

To add tags to an existing SSM document, use the AddTagsToResource operation.

content :: Text

The content for the new SSM document in JSON or YAML format. We recommend storing the contents for your new document in an external JSON or YAML file and referencing the file in a command.

For examples, see the following topics in the /Amazon Web Services Systems Manager User Guide/.

Create an SSM document (Amazon Web Services API)

Create an SSM document (Amazon Web Services CLI)

Create an SSM document (API)

name :: Text

A name for the SSM document.

You can't use the following strings as document name prefixes. These are reserved by Amazon Web Services for use as document name prefixes:

aws-

amazon

amzn

newCreateDocument

Arguments

:: Text
$sel:content:CreateDocument'

-> Text
$sel:name:CreateDocument'

-> CreateDocument

Create a value of CreateDocument with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:documentType:CreateDocument', createDocument_documentType - The type of document to create.

$sel:attachments:CreateDocument', createDocument_attachments - A list of key-value pairs that describe attachments to a version of a document.

$sel:versionName:CreateDocument', createDocument_versionName - An optional field specifying the version of the artifact you are creating with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

$sel:targetType:CreateDocument', createDocument_targetType - Specify a target type to define the kinds of resources the document can run on. For example, to run a document on EC2 instances, specify the following value: /AWS::EC2::Instance. If you specify a value of '/' the document can run on all types of resources. If you don't specify a value, the document can't run on any resources. For a list of valid resource types, see Amazon Web Services resource and property types reference in the CloudFormation User Guide.

$sel:documentFormat:CreateDocument', createDocument_documentFormat - Specify the document format for the request. The document format can be JSON, YAML, or TEXT. JSON is the default format.

$sel:displayName:CreateDocument', createDocument_displayName - An optional field where you can specify a friendly name for the SSM document. This value can differ for each version of the document. You can update this value at a later time using the UpdateDocument operation.

$sel:requires:CreateDocument', createDocument_requires - A list of SSM documents required by a document. This parameter is used exclusively by AppConfig. When a user creates an AppConfig configuration in an SSM document, the user must also specify a required document for validation purposes. In this case, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document for validation purposes. For more information, see What is AppConfig? in the AppConfig User Guide.

$sel:tags:CreateDocument', createDocument_tags - Optional metadata that you assign to a resource. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag an SSM document to identify the types of targets or the environment where it will run. In this case, you could specify the following key-value pairs:

Key=OS,Value=Windows

Key=Environment,Value=Production

To add tags to an existing SSM document, use the AddTagsToResource operation.

$sel:content:CreateDocument', createDocument_content - The content for the new SSM document in JSON or YAML format. We recommend storing the contents for your new document in an external JSON or YAML file and referencing the file in a command.

For examples, see the following topics in the /Amazon Web Services Systems Manager User Guide/.

Create an SSM document (Amazon Web Services API)

Create an SSM document (Amazon Web Services CLI)

Create an SSM document (API)

$sel:name:CreateDocument', createDocument_name - A name for the SSM document.

You can't use the following strings as document name prefixes. These are reserved by Amazon Web Services for use as document name prefixes:

aws-

amazon

amzn

Request Lenses

createDocument_documentType :: Lens' CreateDocument (Maybe DocumentType)

The type of document to create.

createDocument_attachments :: Lens' CreateDocument (Maybe [AttachmentsSource])

A list of key-value pairs that describe attachments to a version of a document.

createDocument_versionName :: Lens' CreateDocument (Maybe Text)

An optional field specifying the version of the artifact you are creating with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

createDocument_targetType :: Lens' CreateDocument (Maybe Text)

Specify a target type to define the kinds of resources the document can run on. For example, to run a document on EC2 instances, specify the following value: /AWS::EC2::Instance. If you specify a value of '/' the document can run on all types of resources. If you don't specify a value, the document can't run on any resources. For a list of valid resource types, see Amazon Web Services resource and property types reference in the CloudFormation User Guide.

createDocument_documentFormat :: Lens' CreateDocument (Maybe DocumentFormat)

Specify the document format for the request. The document format can be JSON, YAML, or TEXT. JSON is the default format.

createDocument_displayName :: Lens' CreateDocument (Maybe Text)

An optional field where you can specify a friendly name for the SSM document. This value can differ for each version of the document. You can update this value at a later time using the UpdateDocument operation.

createDocument_requires :: Lens' CreateDocument (Maybe (NonEmpty DocumentRequires))

A list of SSM documents required by a document. This parameter is used exclusively by AppConfig. When a user creates an AppConfig configuration in an SSM document, the user must also specify a required document for validation purposes. In this case, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document for validation purposes. For more information, see What is AppConfig? in the AppConfig User Guide.

createDocument_tags :: Lens' CreateDocument (Maybe [Tag])

Optional metadata that you assign to a resource. Tags enable you to categorize a resource in different ways, such as by purpose, owner, or environment. For example, you might want to tag an SSM document to identify the types of targets or the environment where it will run. In this case, you could specify the following key-value pairs:

Key=OS,Value=Windows

Key=Environment,Value=Production

To add tags to an existing SSM document, use the AddTagsToResource operation.

createDocument_content :: Lens' CreateDocument Text

The content for the new SSM document in JSON or YAML format. We recommend storing the contents for your new document in an external JSON or YAML file and referencing the file in a command.

For examples, see the following topics in the /Amazon Web Services Systems Manager User Guide/.

Create an SSM document (Amazon Web Services API)

Create an SSM document (Amazon Web Services CLI)

Create an SSM document (API)

createDocument_name :: Lens' CreateDocument Text

A name for the SSM document.

You can't use the following strings as document name prefixes. These are reserved by Amazon Web Services for use as document name prefixes:

aws-

amazon

amzn

Destructuring the Response

data CreateDocumentResponse

See: newCreateDocumentResponse smart constructor.

Constructors

CreateDocumentResponse'

Fields

documentDescription :: Maybe DocumentDescription

Information about the SSM document.

httpStatus :: Int

The response's http status code.

newCreateDocumentResponse

Arguments

:: Int
$sel:httpStatus:CreateDocumentResponse'

-> CreateDocumentResponse

Create a value of CreateDocumentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:documentDescription:CreateDocumentResponse', createDocumentResponse_documentDescription - Information about the SSM document.

$sel:httpStatus:CreateDocumentResponse', createDocumentResponse_httpStatus - The response's http status code.

Response Lenses

createDocumentResponse_documentDescription :: Lens' CreateDocumentResponse (Maybe DocumentDescription)

Information about the SSM document.

createDocumentResponse_httpStatus :: Lens' CreateDocumentResponse Int

The response's http status code.