Amazonka.SSM.CreateResourceDataSync

Description

A resource data sync helps you view data from multiple sources in a single location. Amazon Web Services Systems Manager offers two types of resource data sync: SyncToDestination and SyncFromSource.

You can configure Systems Manager Inventory to use the SyncToDestination type to synchronize Inventory data from multiple Amazon Web Services Regions to a single Amazon Simple Storage Service (Amazon S3) bucket. For more information, see Configuring resource data sync for Inventory in the Amazon Web Services Systems Manager User Guide.

You can configure Systems Manager Explorer to use the SyncFromSource type to synchronize operational work items (OpsItems) and operational data (OpsData) from multiple Amazon Web Services Regions to a single Amazon S3 bucket. This type can synchronize OpsItems and OpsData from multiple Amazon Web Services accounts and Amazon Web Services Regions or EntireOrganization by using Organizations. For more information, see Setting up Systems Manager Explorer to display data from multiple accounts and Regions in the Amazon Web Services Systems Manager User Guide.

A resource data sync is an asynchronous operation that returns immediately. After a successful initial sync is completed, the system continuously syncs data. To check the status of a sync, use the ListResourceDataSync.

By default, data isn't encrypted in Amazon S3. We strongly recommend that you enable encryption in Amazon S3 to ensure secure data storage. We also recommend that you secure access to the Amazon S3 bucket by creating a restrictive bucket policy.

Creating a Request

data CreateResourceDataSync

See: newCreateResourceDataSync smart constructor.

Constructors

CreateResourceDataSync'

Fields

syncType :: Maybe Text

Specify SyncToDestination to create a resource data sync that synchronizes data to an S3 bucket for Inventory. If you specify SyncToDestination, you must provide a value for S3Destination. Specify SyncFromSource to synchronize data from a single account and multiple Regions, or multiple Amazon Web Services accounts and Amazon Web Services Regions, as listed in Organizations for Explorer. If you specify SyncFromSource, you must provide a value for SyncSource. The default value is SyncToDestination.

syncSource :: Maybe ResourceDataSyncSource

Specify information about the data sources to synchronize. This parameter is required if the SyncType value is SyncFromSource.

s3Destination :: Maybe ResourceDataSyncS3Destination

Amazon S3 configuration details for the sync. This parameter is required if the SyncType value is SyncToDestination.

syncName :: Text

A name for the configuration.

newCreateResourceDataSync

Arguments

:: Text
$sel:syncName:CreateResourceDataSync'

-> CreateResourceDataSync

Create a value of CreateResourceDataSync with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:syncType:CreateResourceDataSync', createResourceDataSync_syncType - Specify SyncToDestination to create a resource data sync that synchronizes data to an S3 bucket for Inventory. If you specify SyncToDestination, you must provide a value for S3Destination. Specify SyncFromSource to synchronize data from a single account and multiple Regions, or multiple Amazon Web Services accounts and Amazon Web Services Regions, as listed in Organizations for Explorer. If you specify SyncFromSource, you must provide a value for SyncSource. The default value is SyncToDestination.

$sel:syncSource:CreateResourceDataSync', createResourceDataSync_syncSource - Specify information about the data sources to synchronize. This parameter is required if the SyncType value is SyncFromSource.

$sel:s3Destination:CreateResourceDataSync', createResourceDataSync_s3Destination - Amazon S3 configuration details for the sync. This parameter is required if the SyncType value is SyncToDestination.

$sel:syncName:CreateResourceDataSync', createResourceDataSync_syncName - A name for the configuration.

Request Lenses

createResourceDataSync_syncType :: Lens' CreateResourceDataSync (Maybe Text)

Specify SyncToDestination to create a resource data sync that synchronizes data to an S3 bucket for Inventory. If you specify SyncToDestination, you must provide a value for S3Destination. Specify SyncFromSource to synchronize data from a single account and multiple Regions, or multiple Amazon Web Services accounts and Amazon Web Services Regions, as listed in Organizations for Explorer. If you specify SyncFromSource, you must provide a value for SyncSource. The default value is SyncToDestination.

createResourceDataSync_syncSource :: Lens' CreateResourceDataSync (Maybe ResourceDataSyncSource)

Specify information about the data sources to synchronize. This parameter is required if the SyncType value is SyncFromSource.

createResourceDataSync_s3Destination :: Lens' CreateResourceDataSync (Maybe ResourceDataSyncS3Destination)

Amazon S3 configuration details for the sync. This parameter is required if the SyncType value is SyncToDestination.

createResourceDataSync_syncName :: Lens' CreateResourceDataSync Text

A name for the configuration.

Destructuring the Response

data CreateResourceDataSyncResponse

See: newCreateResourceDataSyncResponse smart constructor.

Constructors

CreateResourceDataSyncResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateResourceDataSyncResponse

Arguments

:: Int
$sel:httpStatus:CreateResourceDataSyncResponse'

-> CreateResourceDataSyncResponse

Create a value of CreateResourceDataSyncResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateResourceDataSyncResponse', createResourceDataSyncResponse_httpStatus - The response's http status code.

Response Lenses

createResourceDataSyncResponse_httpStatus :: Lens' CreateResourceDataSyncResponse Int

The response's http status code.