Amazonka.SSM.DeregisterPatchBaselineForPatchGroup

Description

Removes a patch group from a patch baseline.

Creating a Request

data DeregisterPatchBaselineForPatchGroup

See: newDeregisterPatchBaselineForPatchGroup smart constructor.

Constructors

DeregisterPatchBaselineForPatchGroup'

Fields

baselineId :: Text

The ID of the patch baseline to deregister the patch group from.

patchGroup :: Text

The name of the patch group that should be deregistered from the patch baseline.

newDeregisterPatchBaselineForPatchGroup

Arguments

:: Text
$sel:baselineId:DeregisterPatchBaselineForPatchGroup'

-> Text
$sel:patchGroup:DeregisterPatchBaselineForPatchGroup'

-> DeregisterPatchBaselineForPatchGroup

Create a value of DeregisterPatchBaselineForPatchGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baselineId:DeregisterPatchBaselineForPatchGroup', deregisterPatchBaselineForPatchGroup_baselineId - The ID of the patch baseline to deregister the patch group from.

$sel:patchGroup:DeregisterPatchBaselineForPatchGroup', deregisterPatchBaselineForPatchGroup_patchGroup - The name of the patch group that should be deregistered from the patch baseline.

Request Lenses

deregisterPatchBaselineForPatchGroup_baselineId :: Lens' DeregisterPatchBaselineForPatchGroup Text

The ID of the patch baseline to deregister the patch group from.

deregisterPatchBaselineForPatchGroup_patchGroup :: Lens' DeregisterPatchBaselineForPatchGroup Text

The name of the patch group that should be deregistered from the patch baseline.

Destructuring the Response

data DeregisterPatchBaselineForPatchGroupResponse

See: newDeregisterPatchBaselineForPatchGroupResponse smart constructor.

Constructors

DeregisterPatchBaselineForPatchGroupResponse'

Fields

baselineId :: Maybe Text

The ID of the patch baseline the patch group was deregistered from.

patchGroup :: Maybe Text

The name of the patch group deregistered from the patch baseline.

httpStatus :: Int

The response's http status code.

newDeregisterPatchBaselineForPatchGroupResponse

Arguments

:: Int
$sel:httpStatus:DeregisterPatchBaselineForPatchGroupResponse'

-> DeregisterPatchBaselineForPatchGroupResponse

Create a value of DeregisterPatchBaselineForPatchGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baselineId:DeregisterPatchBaselineForPatchGroup', deregisterPatchBaselineForPatchGroupResponse_baselineId - The ID of the patch baseline the patch group was deregistered from.

$sel:patchGroup:DeregisterPatchBaselineForPatchGroup', deregisterPatchBaselineForPatchGroupResponse_patchGroup - The name of the patch group deregistered from the patch baseline.

$sel:httpStatus:DeregisterPatchBaselineForPatchGroupResponse', deregisterPatchBaselineForPatchGroupResponse_httpStatus - The response's http status code.

Response Lenses

deregisterPatchBaselineForPatchGroupResponse_baselineId :: Lens' DeregisterPatchBaselineForPatchGroupResponse (Maybe Text)

The ID of the patch baseline the patch group was deregistered from.

deregisterPatchBaselineForPatchGroupResponse_patchGroup :: Lens' DeregisterPatchBaselineForPatchGroupResponse (Maybe Text)

The name of the patch group deregistered from the patch baseline.

deregisterPatchBaselineForPatchGroupResponse_httpStatus :: Lens' DeregisterPatchBaselineForPatchGroupResponse Int

The response's http status code.