Amazonka.SSM.DescribeAssociationExecutionTargets

Description

Views information about a specific execution of a specific association.

This operation returns paginated results.

Creating a Request

data DescribeAssociationExecutionTargets

See: newDescribeAssociationExecutionTargets smart constructor.

Constructors

DescribeAssociationExecutionTargets'

Fields

filters :: Maybe (NonEmpty AssociationExecutionTargetsFilter)

Filters for the request. You can specify the following filters and values.

Status (EQUAL)

ResourceId (EQUAL)

ResourceType (EQUAL)

nextToken :: Maybe Text

A token to start the list. Use this token to get the next set of results.

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

associationId :: Text

The association ID that includes the execution for which you want to view details.

executionId :: Text

The execution ID for which you want to view details.

newDescribeAssociationExecutionTargets

Arguments

:: Text
$sel:associationId:DescribeAssociationExecutionTargets'

-> Text
$sel:executionId:DescribeAssociationExecutionTargets'

-> DescribeAssociationExecutionTargets

Create a value of DescribeAssociationExecutionTargets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeAssociationExecutionTargets', describeAssociationExecutionTargets_filters - Filters for the request. You can specify the following filters and values.

Status (EQUAL)

ResourceId (EQUAL)

ResourceType (EQUAL)

$sel:nextToken:DescribeAssociationExecutionTargets', describeAssociationExecutionTargets_nextToken - A token to start the list. Use this token to get the next set of results.

$sel:maxResults:DescribeAssociationExecutionTargets', describeAssociationExecutionTargets_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:associationId:DescribeAssociationExecutionTargets', describeAssociationExecutionTargets_associationId - The association ID that includes the execution for which you want to view details.

$sel:executionId:DescribeAssociationExecutionTargets', describeAssociationExecutionTargets_executionId - The execution ID for which you want to view details.

Request Lenses

describeAssociationExecutionTargets_filters :: Lens' DescribeAssociationExecutionTargets (Maybe (NonEmpty AssociationExecutionTargetsFilter))

Filters for the request. You can specify the following filters and values.

Status (EQUAL)

ResourceId (EQUAL)

ResourceType (EQUAL)

describeAssociationExecutionTargets_nextToken :: Lens' DescribeAssociationExecutionTargets (Maybe Text)

A token to start the list. Use this token to get the next set of results.

describeAssociationExecutionTargets_maxResults :: Lens' DescribeAssociationExecutionTargets (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

describeAssociationExecutionTargets_associationId :: Lens' DescribeAssociationExecutionTargets Text

The association ID that includes the execution for which you want to view details.

describeAssociationExecutionTargets_executionId :: Lens' DescribeAssociationExecutionTargets Text

The execution ID for which you want to view details.

Destructuring the Response

data DescribeAssociationExecutionTargetsResponse

See: newDescribeAssociationExecutionTargetsResponse smart constructor.

Constructors

DescribeAssociationExecutionTargetsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. Use this token to get the next set of results.

associationExecutionTargets :: Maybe [AssociationExecutionTarget]

Information about the execution.

httpStatus :: Int

The response's http status code.

newDescribeAssociationExecutionTargetsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAssociationExecutionTargetsResponse'

-> DescribeAssociationExecutionTargetsResponse

Create a value of DescribeAssociationExecutionTargetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeAssociationExecutionTargets', describeAssociationExecutionTargetsResponse_nextToken - The token for the next set of items to return. Use this token to get the next set of results.

$sel:associationExecutionTargets:DescribeAssociationExecutionTargetsResponse', describeAssociationExecutionTargetsResponse_associationExecutionTargets - Information about the execution.

$sel:httpStatus:DescribeAssociationExecutionTargetsResponse', describeAssociationExecutionTargetsResponse_httpStatus - The response's http status code.

Response Lenses

describeAssociationExecutionTargetsResponse_nextToken :: Lens' DescribeAssociationExecutionTargetsResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

describeAssociationExecutionTargetsResponse_associationExecutionTargets :: Lens' DescribeAssociationExecutionTargetsResponse (Maybe [AssociationExecutionTarget])

Information about the execution.

describeAssociationExecutionTargetsResponse_httpStatus :: Lens' DescribeAssociationExecutionTargetsResponse Int

The response's http status code.