Amazonka.SSM.DescribeAutomationStepExecutions

Description

Information about all active and terminated step executions in an Automation workflow.

This operation returns paginated results.

Creating a Request

data DescribeAutomationStepExecutions

See: newDescribeAutomationStepExecutions smart constructor.

Constructors

DescribeAutomationStepExecutions'

Fields

filters :: Maybe (NonEmpty StepExecutionFilter)

One or more filters to limit the number of step executions returned by the request.

reverseOrder :: Maybe Bool

Indicates whether to list step executions in reverse order by start time. The default value is 'false'.

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

automationExecutionId :: Text

The Automation execution ID for which you want step execution descriptions.

newDescribeAutomationStepExecutions

Arguments

:: Text
$sel:automationExecutionId:DescribeAutomationStepExecutions'

-> DescribeAutomationStepExecutions

Create a value of DescribeAutomationStepExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeAutomationStepExecutions', describeAutomationStepExecutions_filters - One or more filters to limit the number of step executions returned by the request.

$sel:reverseOrder:DescribeAutomationStepExecutions', describeAutomationStepExecutions_reverseOrder - Indicates whether to list step executions in reverse order by start time. The default value is 'false'.

$sel:nextToken:DescribeAutomationStepExecutions', describeAutomationStepExecutions_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:DescribeAutomationStepExecutions', describeAutomationStepExecutions_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:automationExecutionId:DescribeAutomationStepExecutions', describeAutomationStepExecutions_automationExecutionId - The Automation execution ID for which you want step execution descriptions.

Request Lenses

describeAutomationStepExecutions_filters :: Lens' DescribeAutomationStepExecutions (Maybe (NonEmpty StepExecutionFilter))

One or more filters to limit the number of step executions returned by the request.

describeAutomationStepExecutions_reverseOrder :: Lens' DescribeAutomationStepExecutions (Maybe Bool)

Indicates whether to list step executions in reverse order by start time. The default value is 'false'.

describeAutomationStepExecutions_nextToken :: Lens' DescribeAutomationStepExecutions (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeAutomationStepExecutions_maxResults :: Lens' DescribeAutomationStepExecutions (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

describeAutomationStepExecutions_automationExecutionId :: Lens' DescribeAutomationStepExecutions Text

The Automation execution ID for which you want step execution descriptions.

Destructuring the Response

data DescribeAutomationStepExecutionsResponse

See: newDescribeAutomationStepExecutionsResponse smart constructor.

Constructors

DescribeAutomationStepExecutionsResponse'

Fields

nextToken :: Maybe Text

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

stepExecutions :: Maybe [StepExecution]

A list of details about the current state of all steps that make up an execution.

httpStatus :: Int

The response's http status code.

newDescribeAutomationStepExecutionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAutomationStepExecutionsResponse'

-> DescribeAutomationStepExecutionsResponse

Create a value of DescribeAutomationStepExecutionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeAutomationStepExecutions', describeAutomationStepExecutionsResponse_nextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

$sel:stepExecutions:DescribeAutomationStepExecutionsResponse', describeAutomationStepExecutionsResponse_stepExecutions - A list of details about the current state of all steps that make up an execution.

$sel:httpStatus:DescribeAutomationStepExecutionsResponse', describeAutomationStepExecutionsResponse_httpStatus - The response's http status code.

Response Lenses

describeAutomationStepExecutionsResponse_nextToken :: Lens' DescribeAutomationStepExecutionsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

describeAutomationStepExecutionsResponse_stepExecutions :: Lens' DescribeAutomationStepExecutionsResponse (Maybe [StepExecution])

A list of details about the current state of all steps that make up an execution.

describeAutomationStepExecutionsResponse_httpStatus :: Lens' DescribeAutomationStepExecutionsResponse Int

The response's http status code.