Amazonka.SSM.DescribeMaintenanceWindowTargets

Description

Lists the targets registered with the maintenance window.

This operation returns paginated results.

Creating a Request

data DescribeMaintenanceWindowTargets

See: newDescribeMaintenanceWindowTargets smart constructor.

Constructors

DescribeMaintenanceWindowTargets'

Fields

filters :: Maybe [MaintenanceWindowFilter]

Optional filters that can be used to narrow down the scope of the returned window targets. The supported filter keys are Type, WindowTargetId, and OwnerInformation.

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

windowId :: Text

The ID of the maintenance window whose targets should be retrieved.

newDescribeMaintenanceWindowTargets

Arguments

:: Text
$sel:windowId:DescribeMaintenanceWindowTargets'

-> DescribeMaintenanceWindowTargets

Create a value of DescribeMaintenanceWindowTargets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeMaintenanceWindowTargets', describeMaintenanceWindowTargets_filters - Optional filters that can be used to narrow down the scope of the returned window targets. The supported filter keys are Type, WindowTargetId, and OwnerInformation.

$sel:nextToken:DescribeMaintenanceWindowTargets', describeMaintenanceWindowTargets_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:DescribeMaintenanceWindowTargets', describeMaintenanceWindowTargets_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:windowId:DescribeMaintenanceWindowTargets', describeMaintenanceWindowTargets_windowId - The ID of the maintenance window whose targets should be retrieved.

Request Lenses

describeMaintenanceWindowTargets_filters :: Lens' DescribeMaintenanceWindowTargets (Maybe [MaintenanceWindowFilter])

Optional filters that can be used to narrow down the scope of the returned window targets. The supported filter keys are Type, WindowTargetId, and OwnerInformation.

describeMaintenanceWindowTargets_nextToken :: Lens' DescribeMaintenanceWindowTargets (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeMaintenanceWindowTargets_maxResults :: Lens' DescribeMaintenanceWindowTargets (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

describeMaintenanceWindowTargets_windowId :: Lens' DescribeMaintenanceWindowTargets Text

The ID of the maintenance window whose targets should be retrieved.

Destructuring the Response

data DescribeMaintenanceWindowTargetsResponse

See: newDescribeMaintenanceWindowTargetsResponse smart constructor.

Constructors

DescribeMaintenanceWindowTargetsResponse'

Fields

nextToken :: Maybe Text

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

targets :: Maybe [MaintenanceWindowTarget]

Information about the targets in the maintenance window.

httpStatus :: Int

The response's http status code.

newDescribeMaintenanceWindowTargetsResponse

Arguments

:: Int
$sel:httpStatus:DescribeMaintenanceWindowTargetsResponse'

-> DescribeMaintenanceWindowTargetsResponse

Create a value of DescribeMaintenanceWindowTargetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeMaintenanceWindowTargets', describeMaintenanceWindowTargetsResponse_nextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

$sel:targets:DescribeMaintenanceWindowTargetsResponse', describeMaintenanceWindowTargetsResponse_targets - Information about the targets in the maintenance window.

$sel:httpStatus:DescribeMaintenanceWindowTargetsResponse', describeMaintenanceWindowTargetsResponse_httpStatus - The response's http status code.

Response Lenses

describeMaintenanceWindowTargetsResponse_nextToken :: Lens' DescribeMaintenanceWindowTargetsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

describeMaintenanceWindowTargetsResponse_targets :: Lens' DescribeMaintenanceWindowTargetsResponse (Maybe [MaintenanceWindowTarget])

Information about the targets in the maintenance window.

describeMaintenanceWindowTargetsResponse_httpStatus :: Lens' DescribeMaintenanceWindowTargetsResponse Int

The response's http status code.