Amazonka.SSM.GetDocument

Description

Gets the contents of the specified Amazon Web Services Systems Manager document (SSM document).

Creating a Request

data GetDocument

See: newGetDocument smart constructor.

Constructors

GetDocument'

Fields

versionName :: Maybe Text

An optional field specifying the version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document and can't be changed.

documentFormat :: Maybe DocumentFormat

Returns the document in the specified format. The document format can be either JSON or YAML. JSON is the default format.

documentVersion :: Maybe Text

The document version for which you want information.

name :: Text

The name of the SSM document.

newGetDocument

Arguments

:: Text
$sel:name:GetDocument'

-> GetDocument

Create a value of GetDocument with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionName:GetDocument', getDocument_versionName - An optional field specifying the version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document and can't be changed.

$sel:documentFormat:GetDocument', getDocument_documentFormat - Returns the document in the specified format. The document format can be either JSON or YAML. JSON is the default format.

$sel:documentVersion:GetDocument', getDocument_documentVersion - The document version for which you want information.

$sel:name:GetDocument', getDocument_name - The name of the SSM document.

Request Lenses

getDocument_versionName :: Lens' GetDocument (Maybe Text)

An optional field specifying the version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document and can't be changed.

getDocument_documentFormat :: Lens' GetDocument (Maybe DocumentFormat)

Returns the document in the specified format. The document format can be either JSON or YAML. JSON is the default format.

getDocument_documentVersion :: Lens' GetDocument (Maybe Text)

The document version for which you want information.

getDocument_name :: Lens' GetDocument Text

The name of the SSM document.

Destructuring the Response

data GetDocumentResponse

See: newGetDocumentResponse smart constructor.

Constructors

GetDocumentResponse'

Fields

status :: Maybe DocumentStatus

The status of the SSM document, such as Creating, Active, Updating, Failed, and Deleting.

documentType :: Maybe DocumentType

The document type.

versionName :: Maybe Text

The version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

attachmentsContent :: Maybe [AttachmentContent]

A description of the document attachments, including names, locations, sizes, and so on.

reviewStatus :: Maybe ReviewStatus

The current review status of a new custom Systems Manager document (SSM document) created by a member of your organization, or of the latest version of an existing SSM document.

Only one version of an SSM document can be in the APPROVED state at a time. When a new version is approved, the status of the previous version changes to REJECTED.

Only one version of an SSM document can be in review, or PENDING, at a time.

content :: Maybe Text

The contents of the SSM document.

createdDate :: Maybe POSIX

The date the SSM document was created.

documentFormat :: Maybe DocumentFormat

The document format, either JSON or YAML.

name :: Maybe Text

The name of the SSM document.

documentVersion :: Maybe Text

The document version.

displayName :: Maybe Text

The friendly name of the SSM document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

statusInformation :: Maybe Text

A message returned by Amazon Web Services Systems Manager that explains the Status value. For example, a Failed status might be explained by the StatusInformation message, "The specified S3 bucket doesn't exist. Verify that the URL of the S3 bucket is correct."

requires :: Maybe (NonEmpty DocumentRequires)

A list of SSM documents required by a document. For example, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document.

httpStatus :: Int

The response's http status code.

newGetDocumentResponse

Arguments

:: Int
$sel:httpStatus:GetDocumentResponse'

-> GetDocumentResponse

Create a value of GetDocumentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetDocumentResponse', getDocumentResponse_status - The status of the SSM document, such as Creating, Active, Updating, Failed, and Deleting.

$sel:documentType:GetDocumentResponse', getDocumentResponse_documentType - The document type.

$sel:versionName:GetDocument', getDocumentResponse_versionName - The version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

$sel:attachmentsContent:GetDocumentResponse', getDocumentResponse_attachmentsContent - A description of the document attachments, including names, locations, sizes, and so on.

$sel:reviewStatus:GetDocumentResponse', getDocumentResponse_reviewStatus - The current review status of a new custom Systems Manager document (SSM document) created by a member of your organization, or of the latest version of an existing SSM document.

Only one version of an SSM document can be in the APPROVED state at a time. When a new version is approved, the status of the previous version changes to REJECTED.

Only one version of an SSM document can be in review, or PENDING, at a time.

$sel:content:GetDocumentResponse', getDocumentResponse_content - The contents of the SSM document.

$sel:createdDate:GetDocumentResponse', getDocumentResponse_createdDate - The date the SSM document was created.

$sel:documentFormat:GetDocument', getDocumentResponse_documentFormat - The document format, either JSON or YAML.

$sel:name:GetDocument', getDocumentResponse_name - The name of the SSM document.

$sel:documentVersion:GetDocument', getDocumentResponse_documentVersion - The document version.

$sel:displayName:GetDocumentResponse', getDocumentResponse_displayName - The friendly name of the SSM document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

$sel:statusInformation:GetDocumentResponse', getDocumentResponse_statusInformation - A message returned by Amazon Web Services Systems Manager that explains the Status value. For example, a Failed status might be explained by the StatusInformation message, "The specified S3 bucket doesn't exist. Verify that the URL of the S3 bucket is correct."

$sel:requires:GetDocumentResponse', getDocumentResponse_requires - A list of SSM documents required by a document. For example, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document.

$sel:httpStatus:GetDocumentResponse', getDocumentResponse_httpStatus - The response's http status code.

Response Lenses

getDocumentResponse_status :: Lens' GetDocumentResponse (Maybe DocumentStatus)

The status of the SSM document, such as Creating, Active, Updating, Failed, and Deleting.

getDocumentResponse_documentType :: Lens' GetDocumentResponse (Maybe DocumentType)

The document type.

getDocumentResponse_versionName :: Lens' GetDocumentResponse (Maybe Text)

The version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

getDocumentResponse_attachmentsContent :: Lens' GetDocumentResponse (Maybe [AttachmentContent])

A description of the document attachments, including names, locations, sizes, and so on.

getDocumentResponse_reviewStatus :: Lens' GetDocumentResponse (Maybe ReviewStatus)

The current review status of a new custom Systems Manager document (SSM document) created by a member of your organization, or of the latest version of an existing SSM document.

Only one version of an SSM document can be in the APPROVED state at a time. When a new version is approved, the status of the previous version changes to REJECTED.

Only one version of an SSM document can be in review, or PENDING, at a time.

getDocumentResponse_content :: Lens' GetDocumentResponse (Maybe Text)

The contents of the SSM document.

getDocumentResponse_createdDate :: Lens' GetDocumentResponse (Maybe UTCTime)

The date the SSM document was created.

getDocumentResponse_documentFormat :: Lens' GetDocumentResponse (Maybe DocumentFormat)

The document format, either JSON or YAML.

getDocumentResponse_name :: Lens' GetDocumentResponse (Maybe Text)

The name of the SSM document.

getDocumentResponse_documentVersion :: Lens' GetDocumentResponse (Maybe Text)

The document version.

getDocumentResponse_displayName :: Lens' GetDocumentResponse (Maybe Text)

The friendly name of the SSM document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

getDocumentResponse_statusInformation :: Lens' GetDocumentResponse (Maybe Text)

A message returned by Amazon Web Services Systems Manager that explains the Status value. For example, a Failed status might be explained by the StatusInformation message, "The specified S3 bucket doesn't exist. Verify that the URL of the S3 bucket is correct."

getDocumentResponse_requires :: Lens' GetDocumentResponse (Maybe (NonEmpty DocumentRequires))

A list of SSM documents required by a document. For example, an ApplicationConfiguration document requires an ApplicationConfigurationSchema document.

getDocumentResponse_httpStatus :: Lens' GetDocumentResponse Int

The response's http status code.