Amazonka.SSM.GetMaintenanceWindowExecutionTaskInvocation

Description

Retrieves information about a specific task running on a specific target.

Creating a Request

data GetMaintenanceWindowExecutionTaskInvocation

See: newGetMaintenanceWindowExecutionTaskInvocation smart constructor.

Constructors

GetMaintenanceWindowExecutionTaskInvocation'

Fields

windowExecutionId :: Text

The ID of the maintenance window execution for which the task is a part.

taskId :: Text

The ID of the specific task in the maintenance window task that should be retrieved.

invocationId :: Text

The invocation ID to retrieve.

newGetMaintenanceWindowExecutionTaskInvocation

Arguments

:: Text
$sel:windowExecutionId:GetMaintenanceWindowExecutionTaskInvocation'

-> Text
$sel:taskId:GetMaintenanceWindowExecutionTaskInvocation'

-> Text
$sel:invocationId:GetMaintenanceWindowExecutionTaskInvocation'

-> GetMaintenanceWindowExecutionTaskInvocation

Create a value of GetMaintenanceWindowExecutionTaskInvocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:windowExecutionId:GetMaintenanceWindowExecutionTaskInvocation', getMaintenanceWindowExecutionTaskInvocation_windowExecutionId - The ID of the maintenance window execution for which the task is a part.

$sel:taskId:GetMaintenanceWindowExecutionTaskInvocation', getMaintenanceWindowExecutionTaskInvocation_taskId - The ID of the specific task in the maintenance window task that should be retrieved.

$sel:invocationId:GetMaintenanceWindowExecutionTaskInvocation', getMaintenanceWindowExecutionTaskInvocation_invocationId - The invocation ID to retrieve.

Request Lenses

getMaintenanceWindowExecutionTaskInvocation_windowExecutionId :: Lens' GetMaintenanceWindowExecutionTaskInvocation Text

The ID of the maintenance window execution for which the task is a part.

getMaintenanceWindowExecutionTaskInvocation_taskId :: Lens' GetMaintenanceWindowExecutionTaskInvocation Text

The ID of the specific task in the maintenance window task that should be retrieved.

getMaintenanceWindowExecutionTaskInvocation_invocationId :: Lens' GetMaintenanceWindowExecutionTaskInvocation Text

The invocation ID to retrieve.

Destructuring the Response

data GetMaintenanceWindowExecutionTaskInvocationResponse

See: newGetMaintenanceWindowExecutionTaskInvocationResponse smart constructor.

Constructors

GetMaintenanceWindowExecutionTaskInvocationResponse'

Fields

status :: Maybe MaintenanceWindowExecutionStatus

The task status for an invocation.

executionId :: Maybe Text

The execution ID.

taskExecutionId :: Maybe Text

The task execution ID.

startTime :: Maybe POSIX

The time that the task started running on the target.

invocationId :: Maybe Text

The invocation ID.

ownerInformation :: Maybe (Sensitive Text)

User-provided value to be included in any Amazon CloudWatch Events or Amazon EventBridge events raised while running tasks for these targets in this maintenance window.

taskType :: Maybe MaintenanceWindowTaskType

Retrieves the task type for a maintenance window.

windowTargetId :: Maybe Text

The maintenance window target ID.

windowExecutionId :: Maybe Text

The maintenance window execution ID.

statusDetails :: Maybe Text

The details explaining the status. Details are only available for certain status values.

endTime :: Maybe POSIX

The time that the task finished running on the target.

parameters :: Maybe (Sensitive Text)

The parameters used at the time that the task ran.

httpStatus :: Int

The response's http status code.

newGetMaintenanceWindowExecutionTaskInvocationResponse

Arguments

:: Int
$sel:httpStatus:GetMaintenanceWindowExecutionTaskInvocationResponse'

-> GetMaintenanceWindowExecutionTaskInvocationResponse

Create a value of GetMaintenanceWindowExecutionTaskInvocationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetMaintenanceWindowExecutionTaskInvocationResponse', getMaintenanceWindowExecutionTaskInvocationResponse_status - The task status for an invocation.

$sel:executionId:GetMaintenanceWindowExecutionTaskInvocationResponse', getMaintenanceWindowExecutionTaskInvocationResponse_executionId - The execution ID.

$sel:taskExecutionId:GetMaintenanceWindowExecutionTaskInvocationResponse', getMaintenanceWindowExecutionTaskInvocationResponse_taskExecutionId - The task execution ID.

$sel:startTime:GetMaintenanceWindowExecutionTaskInvocationResponse', getMaintenanceWindowExecutionTaskInvocationResponse_startTime - The time that the task started running on the target.

$sel:invocationId:GetMaintenanceWindowExecutionTaskInvocation', getMaintenanceWindowExecutionTaskInvocationResponse_invocationId - The invocation ID.

$sel:ownerInformation:GetMaintenanceWindowExecutionTaskInvocationResponse', getMaintenanceWindowExecutionTaskInvocationResponse_ownerInformation - User-provided value to be included in any Amazon CloudWatch Events or Amazon EventBridge events raised while running tasks for these targets in this maintenance window.

$sel:taskType:GetMaintenanceWindowExecutionTaskInvocationResponse', getMaintenanceWindowExecutionTaskInvocationResponse_taskType - Retrieves the task type for a maintenance window.

$sel:windowTargetId:GetMaintenanceWindowExecutionTaskInvocationResponse', getMaintenanceWindowExecutionTaskInvocationResponse_windowTargetId - The maintenance window target ID.

$sel:windowExecutionId:GetMaintenanceWindowExecutionTaskInvocation', getMaintenanceWindowExecutionTaskInvocationResponse_windowExecutionId - The maintenance window execution ID.

$sel:statusDetails:GetMaintenanceWindowExecutionTaskInvocationResponse', getMaintenanceWindowExecutionTaskInvocationResponse_statusDetails - The details explaining the status. Details are only available for certain status values.

$sel:endTime:GetMaintenanceWindowExecutionTaskInvocationResponse', getMaintenanceWindowExecutionTaskInvocationResponse_endTime - The time that the task finished running on the target.

$sel:parameters:GetMaintenanceWindowExecutionTaskInvocationResponse', getMaintenanceWindowExecutionTaskInvocationResponse_parameters - The parameters used at the time that the task ran.

$sel:httpStatus:GetMaintenanceWindowExecutionTaskInvocationResponse', getMaintenanceWindowExecutionTaskInvocationResponse_httpStatus - The response's http status code.

Response Lenses

getMaintenanceWindowExecutionTaskInvocationResponse_status :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe MaintenanceWindowExecutionStatus)

The task status for an invocation.

getMaintenanceWindowExecutionTaskInvocationResponse_executionId :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The execution ID.

getMaintenanceWindowExecutionTaskInvocationResponse_taskExecutionId :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The task execution ID.

getMaintenanceWindowExecutionTaskInvocationResponse_startTime :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe UTCTime)

The time that the task started running on the target.

getMaintenanceWindowExecutionTaskInvocationResponse_invocationId :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The invocation ID.

getMaintenanceWindowExecutionTaskInvocationResponse_ownerInformation :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

User-provided value to be included in any Amazon CloudWatch Events or Amazon EventBridge events raised while running tasks for these targets in this maintenance window.

getMaintenanceWindowExecutionTaskInvocationResponse_taskType :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe MaintenanceWindowTaskType)

Retrieves the task type for a maintenance window.

getMaintenanceWindowExecutionTaskInvocationResponse_windowTargetId :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The maintenance window target ID.

getMaintenanceWindowExecutionTaskInvocationResponse_windowExecutionId :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The maintenance window execution ID.

getMaintenanceWindowExecutionTaskInvocationResponse_statusDetails :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The details explaining the status. Details are only available for certain status values.

getMaintenanceWindowExecutionTaskInvocationResponse_endTime :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe UTCTime)

The time that the task finished running on the target.

getMaintenanceWindowExecutionTaskInvocationResponse_parameters :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse (Maybe Text)

The parameters used at the time that the task ran.

getMaintenanceWindowExecutionTaskInvocationResponse_httpStatus :: Lens' GetMaintenanceWindowExecutionTaskInvocationResponse Int

The response's http status code.