Amazonka.SSM.GetOpsSummary

Description

View a summary of operations metadata (OpsData) based on specified filters and aggregators. OpsData can include information about Amazon Web Services Systems Manager OpsCenter operational workitems (OpsItems) as well as information about any Amazon Web Services resource or service configured to report OpsData to Amazon Web Services Systems Manager Explorer.

This operation returns paginated results.

Creating a Request

data GetOpsSummary

See: newGetOpsSummary smart constructor.

Constructors

GetOpsSummary'

Fields

aggregators :: Maybe (NonEmpty OpsAggregator)

Optional aggregators that return counts of OpsData based on one or more expressions.

syncName :: Maybe Text

Specify the name of a resource data sync to get.

filters :: Maybe (NonEmpty OpsFilter)

Optional filters used to scope down the returned OpsData.

resultAttributes :: Maybe (NonEmpty OpsResultAttribute)

The OpsData data type to return.

nextToken :: Maybe Text

A token to start the list. Use this token to get the next set of results.

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

newGetOpsSummary :: GetOpsSummary

Create a value of GetOpsSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aggregators:GetOpsSummary', getOpsSummary_aggregators - Optional aggregators that return counts of OpsData based on one or more expressions.

$sel:syncName:GetOpsSummary', getOpsSummary_syncName - Specify the name of a resource data sync to get.

$sel:filters:GetOpsSummary', getOpsSummary_filters - Optional filters used to scope down the returned OpsData.

$sel:resultAttributes:GetOpsSummary', getOpsSummary_resultAttributes - The OpsData data type to return.

$sel:nextToken:GetOpsSummary', getOpsSummary_nextToken - A token to start the list. Use this token to get the next set of results.

$sel:maxResults:GetOpsSummary', getOpsSummary_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Request Lenses

getOpsSummary_aggregators :: Lens' GetOpsSummary (Maybe (NonEmpty OpsAggregator))

Optional aggregators that return counts of OpsData based on one or more expressions.

getOpsSummary_syncName :: Lens' GetOpsSummary (Maybe Text)

Specify the name of a resource data sync to get.

getOpsSummary_filters :: Lens' GetOpsSummary (Maybe (NonEmpty OpsFilter))

Optional filters used to scope down the returned OpsData.

getOpsSummary_resultAttributes :: Lens' GetOpsSummary (Maybe (NonEmpty OpsResultAttribute))

The OpsData data type to return.

getOpsSummary_nextToken :: Lens' GetOpsSummary (Maybe Text)

A token to start the list. Use this token to get the next set of results.

getOpsSummary_maxResults :: Lens' GetOpsSummary (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

Destructuring the Response

data GetOpsSummaryResponse

See: newGetOpsSummaryResponse smart constructor.

Constructors

GetOpsSummaryResponse'

Fields

entities :: Maybe [OpsEntity]

The list of aggregated details and filtered OpsData.

nextToken :: Maybe Text

The token for the next set of items to return. Use this token to get the next set of results.

httpStatus :: Int

The response's http status code.

newGetOpsSummaryResponse

Arguments

:: Int
$sel:httpStatus:GetOpsSummaryResponse'

-> GetOpsSummaryResponse

Create a value of GetOpsSummaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entities:GetOpsSummaryResponse', getOpsSummaryResponse_entities - The list of aggregated details and filtered OpsData.

$sel:nextToken:GetOpsSummary', getOpsSummaryResponse_nextToken - The token for the next set of items to return. Use this token to get the next set of results.

$sel:httpStatus:GetOpsSummaryResponse', getOpsSummaryResponse_httpStatus - The response's http status code.

Response Lenses

getOpsSummaryResponse_entities :: Lens' GetOpsSummaryResponse (Maybe [OpsEntity])

The list of aggregated details and filtered OpsData.

getOpsSummaryResponse_nextToken :: Lens' GetOpsSummaryResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

getOpsSummaryResponse_httpStatus :: Lens' GetOpsSummaryResponse Int

The response's http status code.