Amazonka.SSM.ListComplianceItems

Description

For a specified resource ID, this API operation returns a list of compliance statuses for different resource types. Currently, you can only specify one resource ID per call. List results depend on the criteria specified in the filter.

This operation returns paginated results.

Creating a Request

data ListComplianceItems

See: newListComplianceItems smart constructor.

Constructors

ListComplianceItems'

Fields

resourceIds :: Maybe (NonEmpty Text)

The ID for the resources from which to get compliance information. Currently, you can only specify one resource ID.

filters :: Maybe [ComplianceStringFilter]

One or more compliance filters. Use a filter to return a more specific list of results.

nextToken :: Maybe Text

A token to start the list. Use this token to get the next set of results.

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

resourceTypes :: Maybe (NonEmpty Text)

The type of resource from which to get compliance information. Currently, the only supported resource type is ManagedInstance.

newListComplianceItems :: ListComplianceItems

Create a value of ListComplianceItems with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceIds:ListComplianceItems', listComplianceItems_resourceIds - The ID for the resources from which to get compliance information. Currently, you can only specify one resource ID.

$sel:filters:ListComplianceItems', listComplianceItems_filters - One or more compliance filters. Use a filter to return a more specific list of results.

$sel:nextToken:ListComplianceItems', listComplianceItems_nextToken - A token to start the list. Use this token to get the next set of results.

$sel:maxResults:ListComplianceItems', listComplianceItems_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:resourceTypes:ListComplianceItems', listComplianceItems_resourceTypes - The type of resource from which to get compliance information. Currently, the only supported resource type is ManagedInstance.

Request Lenses

listComplianceItems_resourceIds :: Lens' ListComplianceItems (Maybe (NonEmpty Text))

The ID for the resources from which to get compliance information. Currently, you can only specify one resource ID.

listComplianceItems_filters :: Lens' ListComplianceItems (Maybe [ComplianceStringFilter])

One or more compliance filters. Use a filter to return a more specific list of results.

listComplianceItems_nextToken :: Lens' ListComplianceItems (Maybe Text)

A token to start the list. Use this token to get the next set of results.

listComplianceItems_maxResults :: Lens' ListComplianceItems (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listComplianceItems_resourceTypes :: Lens' ListComplianceItems (Maybe (NonEmpty Text))

The type of resource from which to get compliance information. Currently, the only supported resource type is ManagedInstance.

Destructuring the Response

data ListComplianceItemsResponse

See: newListComplianceItemsResponse smart constructor.

Constructors

ListComplianceItemsResponse'

Fields

complianceItems :: Maybe [ComplianceItem]

A list of compliance information for the specified resource ID.

nextToken :: Maybe Text

The token for the next set of items to return. Use this token to get the next set of results.

httpStatus :: Int

The response's http status code.

newListComplianceItemsResponse

Arguments

:: Int
$sel:httpStatus:ListComplianceItemsResponse'

-> ListComplianceItemsResponse

Create a value of ListComplianceItemsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:complianceItems:ListComplianceItemsResponse', listComplianceItemsResponse_complianceItems - A list of compliance information for the specified resource ID.

$sel:nextToken:ListComplianceItems', listComplianceItemsResponse_nextToken - The token for the next set of items to return. Use this token to get the next set of results.

$sel:httpStatus:ListComplianceItemsResponse', listComplianceItemsResponse_httpStatus - The response's http status code.

Response Lenses

listComplianceItemsResponse_complianceItems :: Lens' ListComplianceItemsResponse (Maybe [ComplianceItem])

A list of compliance information for the specified resource ID.

listComplianceItemsResponse_nextToken :: Lens' ListComplianceItemsResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

listComplianceItemsResponse_httpStatus :: Lens' ListComplianceItemsResponse Int

The response's http status code.