Amazonka.SSM.ListDocumentMetadataHistory

Description

Information about approval reviews for a version of a change template in Change Manager.

Creating a Request

data ListDocumentMetadataHistory

See: newListDocumentMetadataHistory smart constructor.

Constructors

ListDocumentMetadataHistory'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

documentVersion :: Maybe Text

The version of the change template.

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

name :: Text

The name of the change template.

metadata :: DocumentMetadataEnum

The type of data for which details are being requested. Currently, the only supported value is DocumentReviews.

newListDocumentMetadataHistory

Arguments

:: Text
$sel:name:ListDocumentMetadataHistory'

-> DocumentMetadataEnum
$sel:metadata:ListDocumentMetadataHistory'

-> ListDocumentMetadataHistory

Create a value of ListDocumentMetadataHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDocumentMetadataHistory', listDocumentMetadataHistory_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:documentVersion:ListDocumentMetadataHistory', listDocumentMetadataHistory_documentVersion - The version of the change template.

$sel:maxResults:ListDocumentMetadataHistory', listDocumentMetadataHistory_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:name:ListDocumentMetadataHistory', listDocumentMetadataHistory_name - The name of the change template.

$sel:metadata:ListDocumentMetadataHistory', listDocumentMetadataHistory_metadata - The type of data for which details are being requested. Currently, the only supported value is DocumentReviews.

Request Lenses

listDocumentMetadataHistory_nextToken :: Lens' ListDocumentMetadataHistory (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

listDocumentMetadataHistory_documentVersion :: Lens' ListDocumentMetadataHistory (Maybe Text)

The version of the change template.

listDocumentMetadataHistory_maxResults :: Lens' ListDocumentMetadataHistory (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listDocumentMetadataHistory_name :: Lens' ListDocumentMetadataHistory Text

The name of the change template.

listDocumentMetadataHistory_metadata :: Lens' ListDocumentMetadataHistory DocumentMetadataEnum

The type of data for which details are being requested. Currently, the only supported value is DocumentReviews.

Destructuring the Response

data ListDocumentMetadataHistoryResponse

See: newListDocumentMetadataHistoryResponse smart constructor.

Constructors

ListDocumentMetadataHistoryResponse'

Fields

nextToken :: Maybe Text

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

name :: Maybe Text

The name of the change template.

documentVersion :: Maybe Text

The version of the change template.

author :: Maybe Text

The user ID of the person in the organization who requested the review of the change template.

metadata :: Maybe DocumentMetadataResponseInfo

Information about the response to the change template approval request.

httpStatus :: Int

The response's http status code.

newListDocumentMetadataHistoryResponse

Arguments

:: Int
$sel:httpStatus:ListDocumentMetadataHistoryResponse'

-> ListDocumentMetadataHistoryResponse

Create a value of ListDocumentMetadataHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDocumentMetadataHistory', listDocumentMetadataHistoryResponse_nextToken - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:name:ListDocumentMetadataHistory', listDocumentMetadataHistoryResponse_name - The name of the change template.

$sel:documentVersion:ListDocumentMetadataHistory', listDocumentMetadataHistoryResponse_documentVersion - The version of the change template.

$sel:author:ListDocumentMetadataHistoryResponse', listDocumentMetadataHistoryResponse_author - The user ID of the person in the organization who requested the review of the change template.

$sel:metadata:ListDocumentMetadataHistory', listDocumentMetadataHistoryResponse_metadata - Information about the response to the change template approval request.

$sel:httpStatus:ListDocumentMetadataHistoryResponse', listDocumentMetadataHistoryResponse_httpStatus - The response's http status code.

Response Lenses

listDocumentMetadataHistoryResponse_nextToken :: Lens' ListDocumentMetadataHistoryResponse (Maybe Text)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listDocumentMetadataHistoryResponse_name :: Lens' ListDocumentMetadataHistoryResponse (Maybe Text)

The name of the change template.

listDocumentMetadataHistoryResponse_documentVersion :: Lens' ListDocumentMetadataHistoryResponse (Maybe Text)

The version of the change template.

listDocumentMetadataHistoryResponse_author :: Lens' ListDocumentMetadataHistoryResponse (Maybe Text)

The user ID of the person in the organization who requested the review of the change template.

listDocumentMetadataHistoryResponse_metadata :: Lens' ListDocumentMetadataHistoryResponse (Maybe DocumentMetadataResponseInfo)

Information about the response to the change template approval request.

listDocumentMetadataHistoryResponse_httpStatus :: Lens' ListDocumentMetadataHistoryResponse Int

The response's http status code.