Amazonka.SSM.ListInventoryEntries

Description

A list of inventory items returned by the request.

Creating a Request

data ListInventoryEntries

See: newListInventoryEntries smart constructor.

Constructors

ListInventoryEntries'

Fields

filters :: Maybe (NonEmpty InventoryFilter)

One or more filters. Use a filter to return a more specific list of results.

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

instanceId :: Text

The instance ID for which you want inventory information.

typeName :: Text

The type of inventory item for which you want information.

newListInventoryEntries

Arguments

:: Text
$sel:instanceId:ListInventoryEntries'

-> Text
$sel:typeName:ListInventoryEntries'

-> ListInventoryEntries

Create a value of ListInventoryEntries with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListInventoryEntries', listInventoryEntries_filters - One or more filters. Use a filter to return a more specific list of results.

$sel:nextToken:ListInventoryEntries', listInventoryEntries_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxResults:ListInventoryEntries', listInventoryEntries_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:instanceId:ListInventoryEntries', listInventoryEntries_instanceId - The instance ID for which you want inventory information.

$sel:typeName:ListInventoryEntries', listInventoryEntries_typeName - The type of inventory item for which you want information.

Request Lenses

listInventoryEntries_filters :: Lens' ListInventoryEntries (Maybe (NonEmpty InventoryFilter))

One or more filters. Use a filter to return a more specific list of results.

listInventoryEntries_nextToken :: Lens' ListInventoryEntries (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

listInventoryEntries_maxResults :: Lens' ListInventoryEntries (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listInventoryEntries_instanceId :: Lens' ListInventoryEntries Text

The instance ID for which you want inventory information.

listInventoryEntries_typeName :: Lens' ListInventoryEntries Text

The type of inventory item for which you want information.

Destructuring the Response

data ListInventoryEntriesResponse

See: newListInventoryEntriesResponse smart constructor.

Constructors

ListInventoryEntriesResponse'

Fields

instanceId :: Maybe Text

The instance ID targeted by the request to query inventory information.

typeName :: Maybe Text

The type of inventory item returned by the request.

entries :: Maybe [HashMap Text Text]

A list of inventory items on the instance(s).

schemaVersion :: Maybe Text

The inventory schema version used by the instance(s).

captureTime :: Maybe Text

The time that inventory information was collected for the instance(s).

nextToken :: Maybe Text

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

httpStatus :: Int

The response's http status code.

newListInventoryEntriesResponse

Arguments

:: Int
$sel:httpStatus:ListInventoryEntriesResponse'

-> ListInventoryEntriesResponse

Create a value of ListInventoryEntriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:ListInventoryEntries', listInventoryEntriesResponse_instanceId - The instance ID targeted by the request to query inventory information.

$sel:typeName:ListInventoryEntries', listInventoryEntriesResponse_typeName - The type of inventory item returned by the request.

$sel:entries:ListInventoryEntriesResponse', listInventoryEntriesResponse_entries - A list of inventory items on the instance(s).

$sel:schemaVersion:ListInventoryEntriesResponse', listInventoryEntriesResponse_schemaVersion - The inventory schema version used by the instance(s).

$sel:captureTime:ListInventoryEntriesResponse', listInventoryEntriesResponse_captureTime - The time that inventory information was collected for the instance(s).

$sel:nextToken:ListInventoryEntries', listInventoryEntriesResponse_nextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

$sel:httpStatus:ListInventoryEntriesResponse', listInventoryEntriesResponse_httpStatus - The response's http status code.

Response Lenses

listInventoryEntriesResponse_instanceId :: Lens' ListInventoryEntriesResponse (Maybe Text)

The instance ID targeted by the request to query inventory information.

listInventoryEntriesResponse_typeName :: Lens' ListInventoryEntriesResponse (Maybe Text)

The type of inventory item returned by the request.

listInventoryEntriesResponse_entries :: Lens' ListInventoryEntriesResponse (Maybe [HashMap Text Text])

A list of inventory items on the instance(s).

listInventoryEntriesResponse_schemaVersion :: Lens' ListInventoryEntriesResponse (Maybe Text)

The inventory schema version used by the instance(s).

listInventoryEntriesResponse_captureTime :: Lens' ListInventoryEntriesResponse (Maybe Text)

The time that inventory information was collected for the instance(s).

listInventoryEntriesResponse_nextToken :: Lens' ListInventoryEntriesResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

listInventoryEntriesResponse_httpStatus :: Lens' ListInventoryEntriesResponse Int

The response's http status code.