Amazonka.SSM.RegisterTaskWithMaintenanceWindow

Description

Adds a new task to a maintenance window.

Creating a Request

data RegisterTaskWithMaintenanceWindow

See: newRegisterTaskWithMaintenanceWindow smart constructor.

Constructors

RegisterTaskWithMaintenanceWindow'

Fields

serviceRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM service role for Amazon Web Services Systems Manager to assume when running a maintenance window task. If you do not specify a service role ARN, Systems Manager uses your account's service-linked role. If no service-linked role for Systems Manager exists in your account, it is created when you run RegisterTaskWithMaintenanceWindow.

For more information, see the following topics in the in the /Amazon Web Services Systems Manager User Guide/:

Using service-linked roles for Systems Manager

Should I use a service-linked role or a custom service role to run maintenance window tasks?

taskParameters :: Maybe (Sensitive (HashMap Text (Sensitive MaintenanceWindowTaskParameterValueExpression)))

The parameters that should be passed to the task when it is run.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

priority :: Maybe Natural

The priority of the task in the maintenance window, the lower the number the higher the priority. Tasks in a maintenance window are scheduled in priority order with tasks that have the same priority scheduled in parallel.

clientToken :: Maybe Text

User-provided idempotency token.

cutoffBehavior :: Maybe MaintenanceWindowTaskCutoffBehavior

Indicates whether tasks should continue to run after the cutoff time specified in the maintenance windows is reached.

CONTINUE_TASK: When the cutoff time is reached, any tasks that are running continue. The default value.

CANCEL_TASK:

For Automation, Lambda, Step Functions tasks: When the cutoff time is reached, any task invocations that are already running continue, but no new task invocations are started.

For Run Command tasks: When the cutoff time is reached, the system sends a CancelCommand operation that attempts to cancel the command associated with the task. However, there is no guarantee that the command will be terminated and the underlying process stopped.

The status for tasks that are not completed is TIMED_OUT.

maxErrors :: Maybe Text

The maximum number of errors allowed before this task stops being scheduled.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1. This value doesn't affect the running of your task.

taskInvocationParameters :: Maybe MaintenanceWindowTaskInvocationParameters

The parameters that the task should use during execution. Populate only the fields that match the task type. All other fields should be empty.

name :: Maybe Text

An optional name for the task.

targets :: Maybe [Target]

The targets (either instances or maintenance window targets).

One or more targets must be specified for maintenance window Run Command-type tasks. Depending on the task, targets are optional for other maintenance window task types (Automation, Lambda, and Step Functions). For more information about running tasks that don't specify targets, see Registering maintenance window tasks without targets in the Amazon Web Services Systems Manager User Guide.

Specify instances using the following format:

Key=InstanceIds,Values=<instance-id-1>,<instance-id-2>

Specify maintenance window targets using the following format:

Key=WindowTargetIds,Values=<window-target-id-1>,<window-target-id-2>

loggingInfo :: Maybe LoggingInfo

A structure containing information about an Amazon Simple Storage Service (Amazon S3) bucket to write instance-level logs to.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

description :: Maybe (Sensitive Text)

An optional description for the task.

maxConcurrency :: Maybe Text

The maximum number of targets this task can be run for in parallel.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1. This value doesn't affect the running of your task.

windowId :: Text

The ID of the maintenance window the task should be added to.

taskArn :: Text

The ARN of the task to run.

taskType :: MaintenanceWindowTaskType

The type of task being registered.

newRegisterTaskWithMaintenanceWindow

Arguments

:: Text
$sel:windowId:RegisterTaskWithMaintenanceWindow'

-> Text
$sel:taskArn:RegisterTaskWithMaintenanceWindow'

-> MaintenanceWindowTaskType
$sel:taskType:RegisterTaskWithMaintenanceWindow'

-> RegisterTaskWithMaintenanceWindow

Create a value of RegisterTaskWithMaintenanceWindow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceRoleArn:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_serviceRoleArn - The Amazon Resource Name (ARN) of the IAM service role for Amazon Web Services Systems Manager to assume when running a maintenance window task. If you do not specify a service role ARN, Systems Manager uses your account's service-linked role. If no service-linked role for Systems Manager exists in your account, it is created when you run RegisterTaskWithMaintenanceWindow.

For more information, see the following topics in the in the /Amazon Web Services Systems Manager User Guide/:

Using service-linked roles for Systems Manager

Should I use a service-linked role or a custom service role to run maintenance window tasks?

$sel:taskParameters:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_taskParameters - The parameters that should be passed to the task when it is run.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

$sel:priority:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_priority - The priority of the task in the maintenance window, the lower the number the higher the priority. Tasks in a maintenance window are scheduled in priority order with tasks that have the same priority scheduled in parallel.

$sel:clientToken:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_clientToken - User-provided idempotency token.

$sel:cutoffBehavior:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_cutoffBehavior - Indicates whether tasks should continue to run after the cutoff time specified in the maintenance windows is reached.

CONTINUE_TASK: When the cutoff time is reached, any tasks that are running continue. The default value.

CANCEL_TASK:

For Automation, Lambda, Step Functions tasks: When the cutoff time is reached, any task invocations that are already running continue, but no new task invocations are started.

For Run Command tasks: When the cutoff time is reached, the system sends a CancelCommand operation that attempts to cancel the command associated with the task. However, there is no guarantee that the command will be terminated and the underlying process stopped.

The status for tasks that are not completed is TIMED_OUT.

$sel:maxErrors:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_maxErrors - The maximum number of errors allowed before this task stops being scheduled.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1. This value doesn't affect the running of your task.

$sel:taskInvocationParameters:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_taskInvocationParameters - The parameters that the task should use during execution. Populate only the fields that match the task type. All other fields should be empty.

$sel:name:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_name - An optional name for the task.

$sel:targets:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_targets - The targets (either instances or maintenance window targets).

One or more targets must be specified for maintenance window Run Command-type tasks. Depending on the task, targets are optional for other maintenance window task types (Automation, Lambda, and Step Functions). For more information about running tasks that don't specify targets, see Registering maintenance window tasks without targets in the Amazon Web Services Systems Manager User Guide.

Specify instances using the following format:

Key=InstanceIds,Values=<instance-id-1>,<instance-id-2>

Specify maintenance window targets using the following format:

Key=WindowTargetIds,Values=<window-target-id-1>,<window-target-id-2>

$sel:loggingInfo:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_loggingInfo - A structure containing information about an Amazon Simple Storage Service (Amazon S3) bucket to write instance-level logs to.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

$sel:description:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_description - An optional description for the task.

$sel:maxConcurrency:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_maxConcurrency - The maximum number of targets this task can be run for in parallel.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1. This value doesn't affect the running of your task.

$sel:windowId:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_windowId - The ID of the maintenance window the task should be added to.

$sel:taskArn:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_taskArn - The ARN of the task to run.

$sel:taskType:RegisterTaskWithMaintenanceWindow', registerTaskWithMaintenanceWindow_taskType - The type of task being registered.

Request Lenses

registerTaskWithMaintenanceWindow_serviceRoleArn :: Lens' RegisterTaskWithMaintenanceWindow (Maybe Text)

The Amazon Resource Name (ARN) of the IAM service role for Amazon Web Services Systems Manager to assume when running a maintenance window task. If you do not specify a service role ARN, Systems Manager uses your account's service-linked role. If no service-linked role for Systems Manager exists in your account, it is created when you run RegisterTaskWithMaintenanceWindow.

For more information, see the following topics in the in the /Amazon Web Services Systems Manager User Guide/:

Using service-linked roles for Systems Manager

Should I use a service-linked role or a custom service role to run maintenance window tasks?

registerTaskWithMaintenanceWindow_taskParameters :: Lens' RegisterTaskWithMaintenanceWindow (Maybe (HashMap Text MaintenanceWindowTaskParameterValueExpression))

The parameters that should be passed to the task when it is run.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

registerTaskWithMaintenanceWindow_priority :: Lens' RegisterTaskWithMaintenanceWindow (Maybe Natural)

The priority of the task in the maintenance window, the lower the number the higher the priority. Tasks in a maintenance window are scheduled in priority order with tasks that have the same priority scheduled in parallel.

registerTaskWithMaintenanceWindow_clientToken :: Lens' RegisterTaskWithMaintenanceWindow (Maybe Text)

User-provided idempotency token.

registerTaskWithMaintenanceWindow_cutoffBehavior :: Lens' RegisterTaskWithMaintenanceWindow (Maybe MaintenanceWindowTaskCutoffBehavior)

Indicates whether tasks should continue to run after the cutoff time specified in the maintenance windows is reached.

CONTINUE_TASK: When the cutoff time is reached, any tasks that are running continue. The default value.

CANCEL_TASK:

For Automation, Lambda, Step Functions tasks: When the cutoff time is reached, any task invocations that are already running continue, but no new task invocations are started.

For Run Command tasks: When the cutoff time is reached, the system sends a CancelCommand operation that attempts to cancel the command associated with the task. However, there is no guarantee that the command will be terminated and the underlying process stopped.

The status for tasks that are not completed is TIMED_OUT.

registerTaskWithMaintenanceWindow_maxErrors :: Lens' RegisterTaskWithMaintenanceWindow (Maybe Text)

The maximum number of errors allowed before this task stops being scheduled.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1. This value doesn't affect the running of your task.

registerTaskWithMaintenanceWindow_taskInvocationParameters :: Lens' RegisterTaskWithMaintenanceWindow (Maybe MaintenanceWindowTaskInvocationParameters)

The parameters that the task should use during execution. Populate only the fields that match the task type. All other fields should be empty.

registerTaskWithMaintenanceWindow_name :: Lens' RegisterTaskWithMaintenanceWindow (Maybe Text)

An optional name for the task.

registerTaskWithMaintenanceWindow_targets :: Lens' RegisterTaskWithMaintenanceWindow (Maybe [Target])

The targets (either instances or maintenance window targets).

One or more targets must be specified for maintenance window Run Command-type tasks. Depending on the task, targets are optional for other maintenance window task types (Automation, Lambda, and Step Functions). For more information about running tasks that don't specify targets, see Registering maintenance window tasks without targets in the Amazon Web Services Systems Manager User Guide.

Specify instances using the following format:

Key=InstanceIds,Values=<instance-id-1>,<instance-id-2>

Specify maintenance window targets using the following format:

Key=WindowTargetIds,Values=<window-target-id-1>,<window-target-id-2>

registerTaskWithMaintenanceWindow_loggingInfo :: Lens' RegisterTaskWithMaintenanceWindow (Maybe LoggingInfo)

A structure containing information about an Amazon Simple Storage Service (Amazon S3) bucket to write instance-level logs to.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

registerTaskWithMaintenanceWindow_description :: Lens' RegisterTaskWithMaintenanceWindow (Maybe Text)

An optional description for the task.

registerTaskWithMaintenanceWindow_maxConcurrency :: Lens' RegisterTaskWithMaintenanceWindow (Maybe Text)

The maximum number of targets this task can be run for in parallel.

For maintenance window tasks without a target specified, you can't supply a value for this option. Instead, the system inserts a placeholder value of 1. This value doesn't affect the running of your task.

registerTaskWithMaintenanceWindow_windowId :: Lens' RegisterTaskWithMaintenanceWindow Text

The ID of the maintenance window the task should be added to.

registerTaskWithMaintenanceWindow_taskArn :: Lens' RegisterTaskWithMaintenanceWindow Text

The ARN of the task to run.

registerTaskWithMaintenanceWindow_taskType :: Lens' RegisterTaskWithMaintenanceWindow MaintenanceWindowTaskType

The type of task being registered.

Destructuring the Response

data RegisterTaskWithMaintenanceWindowResponse

See: newRegisterTaskWithMaintenanceWindowResponse smart constructor.

Constructors

RegisterTaskWithMaintenanceWindowResponse'

Fields

windowTaskId :: Maybe Text

The ID of the task in the maintenance window.

httpStatus :: Int

The response's http status code.

newRegisterTaskWithMaintenanceWindowResponse

Arguments

:: Int
$sel:httpStatus:RegisterTaskWithMaintenanceWindowResponse'

-> RegisterTaskWithMaintenanceWindowResponse

Create a value of RegisterTaskWithMaintenanceWindowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:windowTaskId:RegisterTaskWithMaintenanceWindowResponse', registerTaskWithMaintenanceWindowResponse_windowTaskId - The ID of the task in the maintenance window.

$sel:httpStatus:RegisterTaskWithMaintenanceWindowResponse', registerTaskWithMaintenanceWindowResponse_httpStatus - The response's http status code.

Response Lenses

registerTaskWithMaintenanceWindowResponse_windowTaskId :: Lens' RegisterTaskWithMaintenanceWindowResponse (Maybe Text)

The ID of the task in the maintenance window.

registerTaskWithMaintenanceWindowResponse_httpStatus :: Lens' RegisterTaskWithMaintenanceWindowResponse Int

The response's http status code.