Amazonka.SSM.ResetServiceSetting

Description

ServiceSetting is an account-level setting for an Amazon Web Services service. This setting defines how a user interacts with or uses a service or a feature of a service. For example, if an Amazon Web Services service charges money to the account based on feature or service usage, then the Amazon Web Services service team might create a default setting of "false". This means the user can't use this feature unless they change the setting to "true" and intentionally opt in for a paid feature.

Services map a SettingId object to a setting value. Amazon Web Services services teams define the default value for a SettingId. You can't create a new SettingId, but you can overwrite the default value if you have the ssm:UpdateServiceSetting permission for the setting. Use the GetServiceSetting API operation to view the current value. Use the UpdateServiceSetting API operation to change the default setting.

Reset the service setting for the account to the default value as provisioned by the Amazon Web Services service team.

Creating a Request

data ResetServiceSetting

The request body of the ResetServiceSetting API operation.

See: newResetServiceSetting smart constructor.

Constructors

ResetServiceSetting'

Fields

settingId :: Text

The Amazon Resource Name (ARN) of the service setting to reset. The setting ID can be one of the following.

/ssm/automation/customer-script-log-destination

/ssm/automation/customer-script-log-group-name

/ssm/documents/console/public-sharing-permission

/ssm/parameter-store/default-parameter-tier

/ssm/parameter-store/high-throughput-enabled

/ssm/managed-instance/activation-tier

newResetServiceSetting

Arguments

:: Text
$sel:settingId:ResetServiceSetting'

-> ResetServiceSetting

Create a value of ResetServiceSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:settingId:ResetServiceSetting', resetServiceSetting_settingId - The Amazon Resource Name (ARN) of the service setting to reset. The setting ID can be one of the following.

/ssm/automation/customer-script-log-destination

/ssm/automation/customer-script-log-group-name

/ssm/documents/console/public-sharing-permission

/ssm/parameter-store/default-parameter-tier

/ssm/parameter-store/high-throughput-enabled

/ssm/managed-instance/activation-tier

Request Lenses

resetServiceSetting_settingId :: Lens' ResetServiceSetting Text

The Amazon Resource Name (ARN) of the service setting to reset. The setting ID can be one of the following.

/ssm/automation/customer-script-log-destination

/ssm/automation/customer-script-log-group-name

/ssm/documents/console/public-sharing-permission

/ssm/parameter-store/default-parameter-tier

/ssm/parameter-store/high-throughput-enabled

/ssm/managed-instance/activation-tier

Destructuring the Response

data ResetServiceSettingResponse

The result body of the ResetServiceSetting API operation.

See: newResetServiceSettingResponse smart constructor.

Constructors

ResetServiceSettingResponse'

Fields

serviceSetting :: Maybe ServiceSetting

The current, effective service setting after calling the ResetServiceSetting API operation.

httpStatus :: Int

The response's http status code.

newResetServiceSettingResponse

Arguments

:: Int
$sel:httpStatus:ResetServiceSettingResponse'

-> ResetServiceSettingResponse

Create a value of ResetServiceSettingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceSetting:ResetServiceSettingResponse', resetServiceSettingResponse_serviceSetting - The current, effective service setting after calling the ResetServiceSetting API operation.

$sel:httpStatus:ResetServiceSettingResponse', resetServiceSettingResponse_httpStatus - The response's http status code.

Response Lenses

resetServiceSettingResponse_serviceSetting :: Lens' ResetServiceSettingResponse (Maybe ServiceSetting)

The current, effective service setting after calling the ResetServiceSetting API operation.

resetServiceSettingResponse_httpStatus :: Lens' ResetServiceSettingResponse Int

The response's http status code.