Amazonka.SSM.Types.AutomationExecution

Description

Documentation

data AutomationExecution

Detailed information about the current state of an individual Automation execution.

See: newAutomationExecution smart constructor.

Constructors

AutomationExecution'

Fields

scheduledTime :: Maybe POSIX

The date and time the Automation operation is scheduled to start.

associationId :: Maybe Text

The ID of a State Manager association used in the Automation operation.

opsItemId :: Maybe Text

The ID of an OpsItem that is created to represent a Change Manager change request.

currentStepName :: Maybe Text

The name of the step that is currently running.

targetParameterName :: Maybe Text

The parameter name.

targetLocations :: Maybe (NonEmpty TargetLocation)

The combination of Amazon Web Services Regions and/or Amazon Web Services accounts where you want to run the Automation.

progressCounters :: Maybe ProgressCounters

An aggregate of step execution statuses displayed in the Amazon Web Services Systems Manager console for a multi-Region and multi-account Automation execution.

executedBy :: Maybe Text

The Amazon Resource Name (ARN) of the user who ran the automation.

documentName :: Maybe Text

The name of the Automation runbook used during the execution.

executionEndTime :: Maybe POSIX

The time the execution finished.

failureMessage :: Maybe Text

A message describing why an execution has failed, if the status is set to Failed.

automationSubtype :: Maybe AutomationSubtype

The subtype of the Automation operation. Currently, the only supported value is ChangeRequest.

mode :: Maybe ExecutionMode

The automation execution mode.

targetMaps :: Maybe [HashMap Text [Text]]

The specified key-value mapping of document parameters to target resources.

stepExecutionsTruncated :: Maybe Bool

A boolean value that indicates if the response contains the full list of the Automation step executions. If true, use the DescribeAutomationStepExecutions API operation to get the full list of step executions.

automationExecutionStatus :: Maybe AutomationExecutionStatus

The execution status of the Automation.

parentAutomationExecutionId :: Maybe Text

The AutomationExecutionId of the parent automation.

outputs :: Maybe (HashMap Text [Text])

The list of execution outputs as defined in the Automation runbook.

maxErrors :: Maybe Text

The MaxErrors value specified by the user when the execution started.

executionStartTime :: Maybe POSIX

The time the execution started.

currentAction :: Maybe Text

The action of the step that is currently running.

targets :: Maybe [Target]

The specified targets.

resolvedTargets :: Maybe ResolvedTargets

A list of resolved targets in the rate control execution.

parameters :: Maybe (HashMap Text [Text])

The key-value map of execution parameters, which were supplied when calling StartAutomationExecution.

documentVersion :: Maybe Text

The version of the document to use during execution.

automationExecutionId :: Maybe Text

The execution ID.

changeRequestName :: Maybe Text

The name of the Change Manager change request.

stepExecutions :: Maybe [StepExecution]

A list of details about the current state of all steps that comprise an execution. An Automation runbook contains a list of steps that are run in order.

runbooks :: Maybe (NonEmpty Runbook)

Information about the Automation runbooks that are run as part of a runbook workflow.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

maxConcurrency :: Maybe Text

The MaxConcurrency value specified by the user when the execution started.

target :: Maybe Text

The target of the execution.

newAutomationExecution :: AutomationExecution

Create a value of AutomationExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduledTime:AutomationExecution', automationExecution_scheduledTime - The date and time the Automation operation is scheduled to start.

$sel:associationId:AutomationExecution', automationExecution_associationId - The ID of a State Manager association used in the Automation operation.

$sel:opsItemId:AutomationExecution', automationExecution_opsItemId - The ID of an OpsItem that is created to represent a Change Manager change request.

$sel:currentStepName:AutomationExecution', automationExecution_currentStepName - The name of the step that is currently running.

$sel:targetParameterName:AutomationExecution', automationExecution_targetParameterName - The parameter name.

$sel:targetLocations:AutomationExecution', automationExecution_targetLocations - The combination of Amazon Web Services Regions and/or Amazon Web Services accounts where you want to run the Automation.

$sel:progressCounters:AutomationExecution', automationExecution_progressCounters - An aggregate of step execution statuses displayed in the Amazon Web Services Systems Manager console for a multi-Region and multi-account Automation execution.

$sel:executedBy:AutomationExecution', automationExecution_executedBy - The Amazon Resource Name (ARN) of the user who ran the automation.

$sel:documentName:AutomationExecution', automationExecution_documentName - The name of the Automation runbook used during the execution.

$sel:executionEndTime:AutomationExecution', automationExecution_executionEndTime - The time the execution finished.

$sel:failureMessage:AutomationExecution', automationExecution_failureMessage - A message describing why an execution has failed, if the status is set to Failed.

$sel:automationSubtype:AutomationExecution', automationExecution_automationSubtype - The subtype of the Automation operation. Currently, the only supported value is ChangeRequest.

$sel:mode:AutomationExecution', automationExecution_mode - The automation execution mode.

$sel:targetMaps:AutomationExecution', automationExecution_targetMaps - The specified key-value mapping of document parameters to target resources.

$sel:stepExecutionsTruncated:AutomationExecution', automationExecution_stepExecutionsTruncated - A boolean value that indicates if the response contains the full list of the Automation step executions. If true, use the DescribeAutomationStepExecutions API operation to get the full list of step executions.

$sel:automationExecutionStatus:AutomationExecution', automationExecution_automationExecutionStatus - The execution status of the Automation.

$sel:parentAutomationExecutionId:AutomationExecution', automationExecution_parentAutomationExecutionId - The AutomationExecutionId of the parent automation.

$sel:outputs:AutomationExecution', automationExecution_outputs - The list of execution outputs as defined in the Automation runbook.

$sel:maxErrors:AutomationExecution', automationExecution_maxErrors - The MaxErrors value specified by the user when the execution started.

$sel:executionStartTime:AutomationExecution', automationExecution_executionStartTime - The time the execution started.

$sel:currentAction:AutomationExecution', automationExecution_currentAction - The action of the step that is currently running.

$sel:targets:AutomationExecution', automationExecution_targets - The specified targets.

$sel:resolvedTargets:AutomationExecution', automationExecution_resolvedTargets - A list of resolved targets in the rate control execution.

$sel:parameters:AutomationExecution', automationExecution_parameters - The key-value map of execution parameters, which were supplied when calling StartAutomationExecution.

$sel:documentVersion:AutomationExecution', automationExecution_documentVersion - The version of the document to use during execution.

$sel:automationExecutionId:AutomationExecution', automationExecution_automationExecutionId - The execution ID.

$sel:changeRequestName:AutomationExecution', automationExecution_changeRequestName - The name of the Change Manager change request.

$sel:stepExecutions:AutomationExecution', automationExecution_stepExecutions - A list of details about the current state of all steps that comprise an execution. An Automation runbook contains a list of steps that are run in order.

$sel:runbooks:AutomationExecution', automationExecution_runbooks - Information about the Automation runbooks that are run as part of a runbook workflow.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

$sel:maxConcurrency:AutomationExecution', automationExecution_maxConcurrency - The MaxConcurrency value specified by the user when the execution started.

$sel:target:AutomationExecution', automationExecution_target - The target of the execution.

automationExecution_scheduledTime :: Lens' AutomationExecution (Maybe UTCTime)

The date and time the Automation operation is scheduled to start.

automationExecution_associationId :: Lens' AutomationExecution (Maybe Text)

The ID of a State Manager association used in the Automation operation.

automationExecution_opsItemId :: Lens' AutomationExecution (Maybe Text)

The ID of an OpsItem that is created to represent a Change Manager change request.

automationExecution_currentStepName :: Lens' AutomationExecution (Maybe Text)

The name of the step that is currently running.

automationExecution_targetParameterName :: Lens' AutomationExecution (Maybe Text)

The parameter name.

automationExecution_targetLocations :: Lens' AutomationExecution (Maybe (NonEmpty TargetLocation))

The combination of Amazon Web Services Regions and/or Amazon Web Services accounts where you want to run the Automation.

automationExecution_progressCounters :: Lens' AutomationExecution (Maybe ProgressCounters)

An aggregate of step execution statuses displayed in the Amazon Web Services Systems Manager console for a multi-Region and multi-account Automation execution.

automationExecution_executedBy :: Lens' AutomationExecution (Maybe Text)

The Amazon Resource Name (ARN) of the user who ran the automation.

automationExecution_documentName :: Lens' AutomationExecution (Maybe Text)

The name of the Automation runbook used during the execution.

automationExecution_executionEndTime :: Lens' AutomationExecution (Maybe UTCTime)

The time the execution finished.

automationExecution_failureMessage :: Lens' AutomationExecution (Maybe Text)

A message describing why an execution has failed, if the status is set to Failed.

automationExecution_automationSubtype :: Lens' AutomationExecution (Maybe AutomationSubtype)

The subtype of the Automation operation. Currently, the only supported value is ChangeRequest.

automationExecution_mode :: Lens' AutomationExecution (Maybe ExecutionMode)

The automation execution mode.

automationExecution_targetMaps :: Lens' AutomationExecution (Maybe [HashMap Text [Text]])

The specified key-value mapping of document parameters to target resources.

automationExecution_stepExecutionsTruncated :: Lens' AutomationExecution (Maybe Bool)

A boolean value that indicates if the response contains the full list of the Automation step executions. If true, use the DescribeAutomationStepExecutions API operation to get the full list of step executions.

automationExecution_automationExecutionStatus :: Lens' AutomationExecution (Maybe AutomationExecutionStatus)

The execution status of the Automation.

automationExecution_parentAutomationExecutionId :: Lens' AutomationExecution (Maybe Text)

The AutomationExecutionId of the parent automation.

automationExecution_outputs :: Lens' AutomationExecution (Maybe (HashMap Text [Text]))

The list of execution outputs as defined in the Automation runbook.

automationExecution_maxErrors :: Lens' AutomationExecution (Maybe Text)

The MaxErrors value specified by the user when the execution started.

automationExecution_executionStartTime :: Lens' AutomationExecution (Maybe UTCTime)

The time the execution started.

automationExecution_currentAction :: Lens' AutomationExecution (Maybe Text)

The action of the step that is currently running.

automationExecution_targets :: Lens' AutomationExecution (Maybe [Target])

The specified targets.

automationExecution_resolvedTargets :: Lens' AutomationExecution (Maybe ResolvedTargets)

A list of resolved targets in the rate control execution.

automationExecution_parameters :: Lens' AutomationExecution (Maybe (HashMap Text [Text]))

The key-value map of execution parameters, which were supplied when calling StartAutomationExecution.

automationExecution_documentVersion :: Lens' AutomationExecution (Maybe Text)

The version of the document to use during execution.

automationExecution_automationExecutionId :: Lens' AutomationExecution (Maybe Text)

The execution ID.

automationExecution_changeRequestName :: Lens' AutomationExecution (Maybe Text)

The name of the Change Manager change request.

automationExecution_stepExecutions :: Lens' AutomationExecution (Maybe [StepExecution])

A list of details about the current state of all steps that comprise an execution. An Automation runbook contains a list of steps that are run in order.

automationExecution_runbooks :: Lens' AutomationExecution (Maybe (NonEmpty Runbook))

Information about the Automation runbooks that are run as part of a runbook workflow.

The Automation runbooks specified for the runbook workflow can't run until all required approvals for the change request have been received.

automationExecution_maxConcurrency :: Lens' AutomationExecution (Maybe Text)

The MaxConcurrency value specified by the user when the execution started.

automationExecution_target :: Lens' AutomationExecution (Maybe Text)

The target of the execution.