Amazonka.SSM.Types.CreateAssociationBatchRequestEntry

Description

Documentation

data CreateAssociationBatchRequestEntry

Describes the association of a Amazon Web Services Systems Manager document (SSM document) and an instance.

See: newCreateAssociationBatchRequestEntry smart constructor.

Constructors

CreateAssociationBatchRequestEntry'

Fields

instanceId :: Maybe Text

The instance ID.

InstanceId has been deprecated. To specify an instance ID for an association, use the Targets parameter. Requests that include the parameter InstanceID with Systems Manager documents (SSM documents) that use schema version 2.0 or later will fail. In addition, if you use the parameter InstanceId, you can't use the parameters AssociationName, DocumentVersion, MaxErrors, MaxConcurrency, OutputLocation, or ScheduleExpression. To use these parameters, you must use the Targets parameter.

targetLocations :: Maybe (NonEmpty TargetLocation)

Use this action to create an association in multiple Regions and multiple accounts.

applyOnlyAtCronInterval :: Maybe Bool

By default, when you create a new associations, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter isn't supported for rate expressions.

maxErrors :: Maybe Text

The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

scheduleExpression :: Maybe Text

A cron expression that specifies a schedule when the association runs.

outputLocation :: Maybe InstanceAssociationOutputLocation

An S3 bucket where you want to store the results of this request.

syncCompliance :: Maybe AssociationSyncCompliance

The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager, a capability of Amazon Web Services Systems Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

targets :: Maybe [Target]

The instances targeted by the request.

parameters :: Maybe (HashMap Text [Text])

A description of the parameters for a document.

documentVersion :: Maybe Text

The document version.

automationTargetParameterName :: Maybe Text

Specify the target for the association. This target is required for associations that use an Automation runbook and target resources by using rate controls. Automation is a capability of Amazon Web Services Systems Manager.

associationName :: Maybe Text

Specify a descriptive name for the association.

calendarNames :: Maybe [Text]

The names or Amazon Resource Names (ARNs) of the Change Calendar type documents your associations are gated under. The associations only run when that Change Calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

complianceSeverity :: Maybe AssociationComplianceSeverity

The severity level to assign to the association.

maxConcurrency :: Maybe Text

The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

name :: Text

The name of the SSM document that contains the configuration information for the instance. You can specify Command or Automation runbooks.

You can specify Amazon Web Services-predefined documents, documents you created, or a document that is shared with you from another account.

For SSM documents that are shared with you from other Amazon Web Services accounts, you must specify the complete SSM document ARN, in the following format:

arn:aws:ssm:region:account-id:document/document-name

For example:

arn:aws:ssm:us-east-2:12345678912:document/My-Shared-Document

For Amazon Web Services-predefined documents and SSM documents you created in your account, you only need to specify the document name. For example, AWS-ApplyPatchBaseline or My-Document.

newCreateAssociationBatchRequestEntry

Arguments

:: Text
$sel:name:CreateAssociationBatchRequestEntry'

-> CreateAssociationBatchRequestEntry

Create a value of CreateAssociationBatchRequestEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_instanceId - The instance ID.

InstanceId has been deprecated. To specify an instance ID for an association, use the Targets parameter. Requests that include the parameter InstanceID with Systems Manager documents (SSM documents) that use schema version 2.0 or later will fail. In addition, if you use the parameter InstanceId, you can't use the parameters AssociationName, DocumentVersion, MaxErrors, MaxConcurrency, OutputLocation, or ScheduleExpression. To use these parameters, you must use the Targets parameter.

$sel:targetLocations:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_targetLocations - Use this action to create an association in multiple Regions and multiple accounts.

$sel:applyOnlyAtCronInterval:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_applyOnlyAtCronInterval - By default, when you create a new associations, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter isn't supported for rate expressions.

$sel:maxErrors:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_maxErrors - The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

$sel:scheduleExpression:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_scheduleExpression - A cron expression that specifies a schedule when the association runs.

$sel:outputLocation:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_outputLocation - An S3 bucket where you want to store the results of this request.

$sel:syncCompliance:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_syncCompliance - The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager, a capability of Amazon Web Services Systems Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

$sel:targets:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_targets - The instances targeted by the request.

$sel:parameters:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_parameters - A description of the parameters for a document.

$sel:documentVersion:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_documentVersion - The document version.

$sel:automationTargetParameterName:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_automationTargetParameterName - Specify the target for the association. This target is required for associations that use an Automation runbook and target resources by using rate controls. Automation is a capability of Amazon Web Services Systems Manager.

$sel:associationName:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_associationName - Specify a descriptive name for the association.

$sel:calendarNames:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_calendarNames - The names or Amazon Resource Names (ARNs) of the Change Calendar type documents your associations are gated under. The associations only run when that Change Calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

$sel:complianceSeverity:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_complianceSeverity - The severity level to assign to the association.

$sel:maxConcurrency:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_maxConcurrency - The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

$sel:name:CreateAssociationBatchRequestEntry', createAssociationBatchRequestEntry_name - The name of the SSM document that contains the configuration information for the instance. You can specify Command or Automation runbooks.

You can specify Amazon Web Services-predefined documents, documents you created, or a document that is shared with you from another account.

For SSM documents that are shared with you from other Amazon Web Services accounts, you must specify the complete SSM document ARN, in the following format:

arn:aws:ssm:region:account-id:document/document-name

For example:

arn:aws:ssm:us-east-2:12345678912:document/My-Shared-Document

For Amazon Web Services-predefined documents and SSM documents you created in your account, you only need to specify the document name. For example, AWS-ApplyPatchBaseline or My-Document.

createAssociationBatchRequestEntry_instanceId :: Lens' CreateAssociationBatchRequestEntry (Maybe Text)

The instance ID.

InstanceId has been deprecated. To specify an instance ID for an association, use the Targets parameter. Requests that include the parameter InstanceID with Systems Manager documents (SSM documents) that use schema version 2.0 or later will fail. In addition, if you use the parameter InstanceId, you can't use the parameters AssociationName, DocumentVersion, MaxErrors, MaxConcurrency, OutputLocation, or ScheduleExpression. To use these parameters, you must use the Targets parameter.

createAssociationBatchRequestEntry_targetLocations :: Lens' CreateAssociationBatchRequestEntry (Maybe (NonEmpty TargetLocation))

Use this action to create an association in multiple Regions and multiple accounts.

createAssociationBatchRequestEntry_applyOnlyAtCronInterval :: Lens' CreateAssociationBatchRequestEntry (Maybe Bool)

By default, when you create a new associations, the system runs it immediately after it is created and then according to the schedule you specified. Specify this option if you don't want an association to run immediately after you create it. This parameter isn't supported for rate expressions.

createAssociationBatchRequestEntry_maxErrors :: Lens' CreateAssociationBatchRequestEntry (Maybe Text)

The number of errors that are allowed before the system stops sending requests to run the association on additional targets. You can specify either an absolute number of errors, for example 10, or a percentage of the target set, for example 10%. If you specify 3, for example, the system stops sending requests when the fourth error is received. If you specify 0, then the system stops sending requests after the first error is returned. If you run an association on 50 instances and set MaxError to 10%, then the system stops sending the request when the sixth error is received.

Executions that are already running an association when MaxErrors is reached are allowed to complete, but some of these executions may fail as well. If you need to ensure that there won't be more than max-errors failed executions, set MaxConcurrency to 1 so that executions proceed one at a time.

createAssociationBatchRequestEntry_scheduleExpression :: Lens' CreateAssociationBatchRequestEntry (Maybe Text)

A cron expression that specifies a schedule when the association runs.

createAssociationBatchRequestEntry_outputLocation :: Lens' CreateAssociationBatchRequestEntry (Maybe InstanceAssociationOutputLocation)

An S3 bucket where you want to store the results of this request.

createAssociationBatchRequestEntry_syncCompliance :: Lens' CreateAssociationBatchRequestEntry (Maybe AssociationSyncCompliance)

The mode for generating association compliance. You can specify AUTO or MANUAL. In AUTO mode, the system uses the status of the association execution to determine the compliance status. If the association execution runs successfully, then the association is COMPLIANT. If the association execution doesn't run successfully, the association is NON-COMPLIANT.

In MANUAL mode, you must specify the AssociationId as a parameter for the PutComplianceItems API operation. In this case, compliance data isn't managed by State Manager, a capability of Amazon Web Services Systems Manager. It is managed by your direct call to the PutComplianceItems API operation.

By default, all associations use AUTO mode.

createAssociationBatchRequestEntry_targets :: Lens' CreateAssociationBatchRequestEntry (Maybe [Target])

The instances targeted by the request.

createAssociationBatchRequestEntry_parameters :: Lens' CreateAssociationBatchRequestEntry (Maybe (HashMap Text [Text]))

A description of the parameters for a document.

createAssociationBatchRequestEntry_documentVersion :: Lens' CreateAssociationBatchRequestEntry (Maybe Text)

The document version.

createAssociationBatchRequestEntry_automationTargetParameterName :: Lens' CreateAssociationBatchRequestEntry (Maybe Text)

Specify the target for the association. This target is required for associations that use an Automation runbook and target resources by using rate controls. Automation is a capability of Amazon Web Services Systems Manager.

createAssociationBatchRequestEntry_associationName :: Lens' CreateAssociationBatchRequestEntry (Maybe Text)

Specify a descriptive name for the association.

createAssociationBatchRequestEntry_calendarNames :: Lens' CreateAssociationBatchRequestEntry (Maybe [Text])

The names or Amazon Resource Names (ARNs) of the Change Calendar type documents your associations are gated under. The associations only run when that Change Calendar is open. For more information, see Amazon Web Services Systems Manager Change Calendar.

createAssociationBatchRequestEntry_complianceSeverity :: Lens' CreateAssociationBatchRequestEntry (Maybe AssociationComplianceSeverity)

The severity level to assign to the association.

createAssociationBatchRequestEntry_maxConcurrency :: Lens' CreateAssociationBatchRequestEntry (Maybe Text)

The maximum number of targets allowed to run the association at the same time. You can specify a number, for example 10, or a percentage of the target set, for example 10%. The default value is 100%, which means all targets run the association at the same time.

If a new instance starts and attempts to run an association while Systems Manager is running MaxConcurrency associations, the association is allowed to run. During the next association interval, the new instance will process its association within the limit specified for MaxConcurrency.

createAssociationBatchRequestEntry_name :: Lens' CreateAssociationBatchRequestEntry Text

The name of the SSM document that contains the configuration information for the instance. You can specify Command or Automation runbooks.

You can specify Amazon Web Services-predefined documents, documents you created, or a document that is shared with you from another account.

For SSM documents that are shared with you from other Amazon Web Services accounts, you must specify the complete SSM document ARN, in the following format:

arn:aws:ssm:region:account-id:document/document-name

For example:

arn:aws:ssm:us-east-2:12345678912:document/My-Shared-Document

For Amazon Web Services-predefined documents and SSM documents you created in your account, you only need to specify the document name. For example, AWS-ApplyPatchBaseline or My-Document.