Amazonka.SSM.Types.DocumentKeyValuesFilter

Description

Documentation

data DocumentKeyValuesFilter

One or more filters. Use a filter to return a more specific list of documents.

For keys, you can specify one or more tags that have been applied to a document.

You can also use Amazon Web Services-provided keys, some of which have specific allowed values. These keys and their associated values are as follows:

DocumentType

- ApplicationConfiguration

ApplicationConfigurationSchema

Automation

ChangeCalendar

Command

DeploymentStrategy

Package

Policy

Session

Owner

Note that only one Owner can be specified in a request. For example: Key=Owner,Values=Self.

Amazon

Private

Public

Self

ThirdParty

PlatformTypes

- Linux

Windows

Name is another Amazon Web Services-provided key. If you use Name as a key, you can use a name prefix to return a list of documents. For example, in the Amazon Web Services CLI, to return a list of all documents that begin with Te, run the following command:

aws ssm list-documents --filters Key=Name,Values=Te

You can also use the TargetType Amazon Web Services-provided key. For a list of valid resource type values that can be used with this key, see Amazon Web Services resource and property types reference in the CloudFormation User Guide.

If you specify more than two keys, only documents that are identified by all the tags are returned in the results. If you specify more than two values for a key, documents that are identified by any of the values are returned in the results.

To specify a custom key-value pair, use the format Key=tag:tagName,Values=valueName.

For example, if you created a key called region and are using the Amazon Web Services CLI to call the list-documents command:

aws ssm list-documents --filters Key=tag:region,Values=east,west Key=Owner,Values=Self

See: newDocumentKeyValuesFilter smart constructor.

Constructors

DocumentKeyValuesFilter'

Fields

values :: Maybe [Text]

The value for the filter key.

key :: Maybe Text

The name of the filter key.

newDocumentKeyValuesFilter :: DocumentKeyValuesFilter

Create a value of DocumentKeyValuesFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:DocumentKeyValuesFilter', documentKeyValuesFilter_values - The value for the filter key.

$sel:key:DocumentKeyValuesFilter', documentKeyValuesFilter_key - The name of the filter key.

documentKeyValuesFilter_values :: Lens' DocumentKeyValuesFilter (Maybe [Text])

The value for the filter key.

documentKeyValuesFilter_key :: Lens' DocumentKeyValuesFilter (Maybe Text)

The name of the filter key.