Amazonka.SSM.Types.DocumentReviewerResponseSource

Description

Documentation

data DocumentReviewerResponseSource

Information about a reviewer's response to a document review request.

See: newDocumentReviewerResponseSource smart constructor.

Constructors

DocumentReviewerResponseSource'

Fields

reviewer :: Maybe Text

The user in your organization assigned to review a document request.

reviewStatus :: Maybe ReviewStatus

The current review status of a new custom SSM document created by a member of your organization, or of the latest version of an existing SSM document.

Only one version of a document can be in the APPROVED state at a time. When a new version is approved, the status of the previous version changes to REJECTED.

Only one version of a document can be in review, or PENDING, at a time.

updatedTime :: Maybe POSIX

The date and time that a reviewer last updated a response to a document review request.

comment :: Maybe [DocumentReviewCommentSource]

The comment entered by a reviewer as part of their document review response.

createTime :: Maybe POSIX

The date and time that a reviewer entered a response to a document review request.

newDocumentReviewerResponseSource :: DocumentReviewerResponseSource

Create a value of DocumentReviewerResponseSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reviewer:DocumentReviewerResponseSource', documentReviewerResponseSource_reviewer - The user in your organization assigned to review a document request.

$sel:reviewStatus:DocumentReviewerResponseSource', documentReviewerResponseSource_reviewStatus - The current review status of a new custom SSM document created by a member of your organization, or of the latest version of an existing SSM document.

Only one version of a document can be in the APPROVED state at a time. When a new version is approved, the status of the previous version changes to REJECTED.

Only one version of a document can be in review, or PENDING, at a time.

$sel:updatedTime:DocumentReviewerResponseSource', documentReviewerResponseSource_updatedTime - The date and time that a reviewer last updated a response to a document review request.

$sel:comment:DocumentReviewerResponseSource', documentReviewerResponseSource_comment - The comment entered by a reviewer as part of their document review response.

$sel:createTime:DocumentReviewerResponseSource', documentReviewerResponseSource_createTime - The date and time that a reviewer entered a response to a document review request.

documentReviewerResponseSource_reviewer :: Lens' DocumentReviewerResponseSource (Maybe Text)

The user in your organization assigned to review a document request.

documentReviewerResponseSource_reviewStatus :: Lens' DocumentReviewerResponseSource (Maybe ReviewStatus)

The current review status of a new custom SSM document created by a member of your organization, or of the latest version of an existing SSM document.

Only one version of a document can be in the APPROVED state at a time. When a new version is approved, the status of the previous version changes to REJECTED.

Only one version of a document can be in review, or PENDING, at a time.

documentReviewerResponseSource_updatedTime :: Lens' DocumentReviewerResponseSource (Maybe UTCTime)

The date and time that a reviewer last updated a response to a document review request.

documentReviewerResponseSource_comment :: Lens' DocumentReviewerResponseSource (Maybe [DocumentReviewCommentSource])

The comment entered by a reviewer as part of their document review response.

documentReviewerResponseSource_createTime :: Lens' DocumentReviewerResponseSource (Maybe UTCTime)

The date and time that a reviewer entered a response to a document review request.