Amazonka.SSM.Types.DocumentVersionInfo

Description

Documentation

data DocumentVersionInfo

Version information about the document.

See: newDocumentVersionInfo smart constructor.

Constructors

DocumentVersionInfo'

Fields

status :: Maybe DocumentStatus

The status of the SSM document, such as Creating, Active, Failed, and Deleting.

versionName :: Maybe Text

The version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

reviewStatus :: Maybe ReviewStatus

The current status of the approval review for the latest version of the document.

createdDate :: Maybe POSIX

The date the document was created.

documentFormat :: Maybe DocumentFormat

The document format, either JSON or YAML.

name :: Maybe Text

The document name.

documentVersion :: Maybe Text

The document version.

displayName :: Maybe Text

The friendly name of the SSM document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

statusInformation :: Maybe Text

A message returned by Amazon Web Services Systems Manager that explains the Status value. For example, a Failed status might be explained by the StatusInformation message, "The specified S3 bucket doesn't exist. Verify that the URL of the S3 bucket is correct."

isDefaultVersion :: Maybe Bool

An identifier for the default version of the document.

newDocumentVersionInfo :: DocumentVersionInfo

Create a value of DocumentVersionInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DocumentVersionInfo', documentVersionInfo_status - The status of the SSM document, such as Creating, Active, Failed, and Deleting.

$sel:versionName:DocumentVersionInfo', documentVersionInfo_versionName - The version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

$sel:reviewStatus:DocumentVersionInfo', documentVersionInfo_reviewStatus - The current status of the approval review for the latest version of the document.

$sel:createdDate:DocumentVersionInfo', documentVersionInfo_createdDate - The date the document was created.

$sel:documentFormat:DocumentVersionInfo', documentVersionInfo_documentFormat - The document format, either JSON or YAML.

$sel:name:DocumentVersionInfo', documentVersionInfo_name - The document name.

$sel:documentVersion:DocumentVersionInfo', documentVersionInfo_documentVersion - The document version.

$sel:displayName:DocumentVersionInfo', documentVersionInfo_displayName - The friendly name of the SSM document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

$sel:statusInformation:DocumentVersionInfo', documentVersionInfo_statusInformation - A message returned by Amazon Web Services Systems Manager that explains the Status value. For example, a Failed status might be explained by the StatusInformation message, "The specified S3 bucket doesn't exist. Verify that the URL of the S3 bucket is correct."

$sel:isDefaultVersion:DocumentVersionInfo', documentVersionInfo_isDefaultVersion - An identifier for the default version of the document.

documentVersionInfo_status :: Lens' DocumentVersionInfo (Maybe DocumentStatus)

The status of the SSM document, such as Creating, Active, Failed, and Deleting.

documentVersionInfo_versionName :: Lens' DocumentVersionInfo (Maybe Text)

The version of the artifact associated with the document. For example, "Release 12, Update 6". This value is unique across all versions of a document, and can't be changed.

documentVersionInfo_reviewStatus :: Lens' DocumentVersionInfo (Maybe ReviewStatus)

The current status of the approval review for the latest version of the document.

documentVersionInfo_createdDate :: Lens' DocumentVersionInfo (Maybe UTCTime)

The date the document was created.

documentVersionInfo_documentFormat :: Lens' DocumentVersionInfo (Maybe DocumentFormat)

The document format, either JSON or YAML.

documentVersionInfo_name :: Lens' DocumentVersionInfo (Maybe Text)

The document name.

documentVersionInfo_documentVersion :: Lens' DocumentVersionInfo (Maybe Text)

The document version.

documentVersionInfo_displayName :: Lens' DocumentVersionInfo (Maybe Text)

The friendly name of the SSM document. This value can differ for each version of the document. If you want to update this value, see UpdateDocument.

documentVersionInfo_statusInformation :: Lens' DocumentVersionInfo (Maybe Text)

A message returned by Amazon Web Services Systems Manager that explains the Status value. For example, a Failed status might be explained by the StatusInformation message, "The specified S3 bucket doesn't exist. Verify that the URL of the S3 bucket is correct."

documentVersionInfo_isDefaultVersion :: Lens' DocumentVersionInfo (Maybe Bool)

An identifier for the default version of the document.