Amazonka.SSM.Types.InstancePatchState

Description

Documentation

data InstancePatchState

Defines the high-level patch compliance state for a managed instance, providing information about the number of installed, missing, not applicable, and failed patches along with metadata about the operation when this information was gathered for the instance.

See: newInstancePatchState smart constructor.

Constructors

InstancePatchState'

Fields

unreportedNotApplicableCount :: Maybe Int

The number of patches beyond the supported limit of NotApplicableCount that aren't reported by name to Inventory. Inventory is a capability of Amazon Web Services Systems Manager.

otherNonCompliantCount :: Maybe Int

The number of instances with patches installed that are specified as other than Critical or Security but aren't compliant with the patch baseline. The status of these instances is NON_COMPLIANT.

rebootOption :: Maybe RebootOption

Indicates the reboot option specified in the patch baseline.

Reboot options apply to Install operations only. Reboots aren't attempted for Patch Manager Scan operations.

RebootIfNeeded: Patch Manager tries to reboot the instance if it installed any patches, or if any patches are detected with a status of InstalledPendingReboot.

NoReboot: Patch Manager attempts to install missing packages without trying to reboot the system. Patches installed with this option are assigned a status of InstalledPendingReboot. These patches might not be in effect until a reboot is performed.

installedPendingRebootCount :: Maybe Int

The number of patches installed by Patch Manager since the last time the instance was rebooted.

ownerInformation :: Maybe (Sensitive Text)

Placeholder information. This field will always be empty in the current release of the service.

securityNonCompliantCount :: Maybe Int

The number of instances where patches that are specified as Security in a patch advisory aren't installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

installedRejectedCount :: Maybe Int

The number of patches installed on an instance that are specified in a RejectedPatches list. Patches with a status of InstalledRejected were typically installed before they were added to a RejectedPatches list.

If ALLOW_AS_DEPENDENCY is the specified option for RejectedPatchesAction, the value of InstalledRejectedCount will always be 0 (zero).

failedCount :: Maybe Int

The number of patches from the patch baseline that were attempted to be installed during the last patching operation, but failed to install.

installedOtherCount :: Maybe Int

The number of patches not specified in the patch baseline that are installed on the instance.

missingCount :: Maybe Int

The number of patches from the patch baseline that are applicable for the instance but aren't currently installed.

installOverrideList :: Maybe Text

An https URL or an Amazon Simple Storage Service (Amazon S3) path-style URL to a list of patches to be installed. This patch installation list, which you maintain in an S3 bucket in YAML format and specify in the SSM document AWS-RunPatchBaseline, overrides the patches specified by the default patch baseline.

For more information about the InstallOverrideList parameter, see About the AWS-RunPatchBaseline SSM document in the Amazon Web Services Systems Manager User Guide.

criticalNonCompliantCount :: Maybe Int

The number of instances where patches that are specified as Critical for compliance reporting in the patch baseline aren't installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

notApplicableCount :: Maybe Int

The number of patches from the patch baseline that aren't applicable for the instance and therefore aren't installed on the instance. This number may be truncated if the list of patch names is very large. The number of patches beyond this limit are reported in UnreportedNotApplicableCount.

installedCount :: Maybe Int

The number of patches from the patch baseline that are installed on the instance.

lastNoRebootInstallOperationTime :: Maybe POSIX

The time of the last attempt to patch the instance with NoReboot specified as the reboot option.

snapshotId :: Maybe Text

The ID of the patch baseline snapshot used during the patching operation when this compliance data was collected.

instanceId :: Text

The ID of the managed instance the high-level patch compliance information was collected for.

patchGroup :: Text

The name of the patch group the managed instance belongs to.

baselineId :: Text

The ID of the patch baseline used to patch the instance.

operationStartTime :: POSIX

The time the most recent patching operation was started on the instance.

operationEndTime :: POSIX

The time the most recent patching operation completed on the instance.

operation :: PatchOperationType

The type of patching operation that was performed: or

SCAN assesses the patch compliance state.

INSTALL installs missing patches.

newInstancePatchState

Arguments

:: Text
$sel:instanceId:InstancePatchState'

-> Text
$sel:patchGroup:InstancePatchState'

-> Text
$sel:baselineId:InstancePatchState'

-> UTCTime
$sel:operationStartTime:InstancePatchState'

-> UTCTime
$sel:operationEndTime:InstancePatchState'

-> PatchOperationType
$sel:operation:InstancePatchState'

-> InstancePatchState

Create a value of InstancePatchState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unreportedNotApplicableCount:InstancePatchState', instancePatchState_unreportedNotApplicableCount - The number of patches beyond the supported limit of NotApplicableCount that aren't reported by name to Inventory. Inventory is a capability of Amazon Web Services Systems Manager.

$sel:otherNonCompliantCount:InstancePatchState', instancePatchState_otherNonCompliantCount - The number of instances with patches installed that are specified as other than Critical or Security but aren't compliant with the patch baseline. The status of these instances is NON_COMPLIANT.

$sel:rebootOption:InstancePatchState', instancePatchState_rebootOption - Indicates the reboot option specified in the patch baseline.

Reboot options apply to Install operations only. Reboots aren't attempted for Patch Manager Scan operations.

RebootIfNeeded: Patch Manager tries to reboot the instance if it installed any patches, or if any patches are detected with a status of InstalledPendingReboot.

NoReboot: Patch Manager attempts to install missing packages without trying to reboot the system. Patches installed with this option are assigned a status of InstalledPendingReboot. These patches might not be in effect until a reboot is performed.

$sel:installedPendingRebootCount:InstancePatchState', instancePatchState_installedPendingRebootCount - The number of patches installed by Patch Manager since the last time the instance was rebooted.

$sel:ownerInformation:InstancePatchState', instancePatchState_ownerInformation - Placeholder information. This field will always be empty in the current release of the service.

$sel:securityNonCompliantCount:InstancePatchState', instancePatchState_securityNonCompliantCount - The number of instances where patches that are specified as Security in a patch advisory aren't installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

$sel:installedRejectedCount:InstancePatchState', instancePatchState_installedRejectedCount - The number of patches installed on an instance that are specified in a RejectedPatches list. Patches with a status of InstalledRejected were typically installed before they were added to a RejectedPatches list.

If ALLOW_AS_DEPENDENCY is the specified option for RejectedPatchesAction, the value of InstalledRejectedCount will always be 0 (zero).

$sel:failedCount:InstancePatchState', instancePatchState_failedCount - The number of patches from the patch baseline that were attempted to be installed during the last patching operation, but failed to install.

$sel:installedOtherCount:InstancePatchState', instancePatchState_installedOtherCount - The number of patches not specified in the patch baseline that are installed on the instance.

$sel:missingCount:InstancePatchState', instancePatchState_missingCount - The number of patches from the patch baseline that are applicable for the instance but aren't currently installed.

$sel:installOverrideList:InstancePatchState', instancePatchState_installOverrideList - An https URL or an Amazon Simple Storage Service (Amazon S3) path-style URL to a list of patches to be installed. This patch installation list, which you maintain in an S3 bucket in YAML format and specify in the SSM document AWS-RunPatchBaseline, overrides the patches specified by the default patch baseline.

For more information about the InstallOverrideList parameter, see About the AWS-RunPatchBaseline SSM document in the Amazon Web Services Systems Manager User Guide.

$sel:criticalNonCompliantCount:InstancePatchState', instancePatchState_criticalNonCompliantCount - The number of instances where patches that are specified as Critical for compliance reporting in the patch baseline aren't installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

$sel:notApplicableCount:InstancePatchState', instancePatchState_notApplicableCount - The number of patches from the patch baseline that aren't applicable for the instance and therefore aren't installed on the instance. This number may be truncated if the list of patch names is very large. The number of patches beyond this limit are reported in UnreportedNotApplicableCount.

$sel:installedCount:InstancePatchState', instancePatchState_installedCount - The number of patches from the patch baseline that are installed on the instance.

$sel:lastNoRebootInstallOperationTime:InstancePatchState', instancePatchState_lastNoRebootInstallOperationTime - The time of the last attempt to patch the instance with NoReboot specified as the reboot option.

$sel:snapshotId:InstancePatchState', instancePatchState_snapshotId - The ID of the patch baseline snapshot used during the patching operation when this compliance data was collected.

$sel:instanceId:InstancePatchState', instancePatchState_instanceId - The ID of the managed instance the high-level patch compliance information was collected for.

$sel:patchGroup:InstancePatchState', instancePatchState_patchGroup - The name of the patch group the managed instance belongs to.

$sel:baselineId:InstancePatchState', instancePatchState_baselineId - The ID of the patch baseline used to patch the instance.

$sel:operationStartTime:InstancePatchState', instancePatchState_operationStartTime - The time the most recent patching operation was started on the instance.

$sel:operationEndTime:InstancePatchState', instancePatchState_operationEndTime - The time the most recent patching operation completed on the instance.

$sel:operation:InstancePatchState', instancePatchState_operation - The type of patching operation that was performed: or

SCAN assesses the patch compliance state.

INSTALL installs missing patches.

instancePatchState_unreportedNotApplicableCount :: Lens' InstancePatchState (Maybe Int)

The number of patches beyond the supported limit of NotApplicableCount that aren't reported by name to Inventory. Inventory is a capability of Amazon Web Services Systems Manager.

instancePatchState_otherNonCompliantCount :: Lens' InstancePatchState (Maybe Int)

The number of instances with patches installed that are specified as other than Critical or Security but aren't compliant with the patch baseline. The status of these instances is NON_COMPLIANT.

instancePatchState_rebootOption :: Lens' InstancePatchState (Maybe RebootOption)

Indicates the reboot option specified in the patch baseline.

Reboot options apply to Install operations only. Reboots aren't attempted for Patch Manager Scan operations.

RebootIfNeeded: Patch Manager tries to reboot the instance if it installed any patches, or if any patches are detected with a status of InstalledPendingReboot.

NoReboot: Patch Manager attempts to install missing packages without trying to reboot the system. Patches installed with this option are assigned a status of InstalledPendingReboot. These patches might not be in effect until a reboot is performed.

instancePatchState_installedPendingRebootCount :: Lens' InstancePatchState (Maybe Int)

The number of patches installed by Patch Manager since the last time the instance was rebooted.

instancePatchState_ownerInformation :: Lens' InstancePatchState (Maybe Text)

Placeholder information. This field will always be empty in the current release of the service.

instancePatchState_securityNonCompliantCount :: Lens' InstancePatchState (Maybe Int)

The number of instances where patches that are specified as Security in a patch advisory aren't installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

instancePatchState_installedRejectedCount :: Lens' InstancePatchState (Maybe Int)

The number of patches installed on an instance that are specified in a RejectedPatches list. Patches with a status of InstalledRejected were typically installed before they were added to a RejectedPatches list.

If ALLOW_AS_DEPENDENCY is the specified option for RejectedPatchesAction, the value of InstalledRejectedCount will always be 0 (zero).

instancePatchState_failedCount :: Lens' InstancePatchState (Maybe Int)

The number of patches from the patch baseline that were attempted to be installed during the last patching operation, but failed to install.

instancePatchState_installedOtherCount :: Lens' InstancePatchState (Maybe Int)

The number of patches not specified in the patch baseline that are installed on the instance.

instancePatchState_missingCount :: Lens' InstancePatchState (Maybe Int)

The number of patches from the patch baseline that are applicable for the instance but aren't currently installed.

instancePatchState_installOverrideList :: Lens' InstancePatchState (Maybe Text)

An https URL or an Amazon Simple Storage Service (Amazon S3) path-style URL to a list of patches to be installed. This patch installation list, which you maintain in an S3 bucket in YAML format and specify in the SSM document AWS-RunPatchBaseline, overrides the patches specified by the default patch baseline.

For more information about the InstallOverrideList parameter, see About the AWS-RunPatchBaseline SSM document in the Amazon Web Services Systems Manager User Guide.

instancePatchState_criticalNonCompliantCount :: Lens' InstancePatchState (Maybe Int)

The number of instances where patches that are specified as Critical for compliance reporting in the patch baseline aren't installed. These patches might be missing, have failed installation, were rejected, or were installed but awaiting a required instance reboot. The status of these instances is NON_COMPLIANT.

instancePatchState_notApplicableCount :: Lens' InstancePatchState (Maybe Int)

The number of patches from the patch baseline that aren't applicable for the instance and therefore aren't installed on the instance. This number may be truncated if the list of patch names is very large. The number of patches beyond this limit are reported in UnreportedNotApplicableCount.

instancePatchState_installedCount :: Lens' InstancePatchState (Maybe Int)

The number of patches from the patch baseline that are installed on the instance.

instancePatchState_lastNoRebootInstallOperationTime :: Lens' InstancePatchState (Maybe UTCTime)

The time of the last attempt to patch the instance with NoReboot specified as the reboot option.

instancePatchState_snapshotId :: Lens' InstancePatchState (Maybe Text)

The ID of the patch baseline snapshot used during the patching operation when this compliance data was collected.

instancePatchState_instanceId :: Lens' InstancePatchState Text

The ID of the managed instance the high-level patch compliance information was collected for.

instancePatchState_patchGroup :: Lens' InstancePatchState Text

The name of the patch group the managed instance belongs to.

instancePatchState_baselineId :: Lens' InstancePatchState Text

The ID of the patch baseline used to patch the instance.

instancePatchState_operationStartTime :: Lens' InstancePatchState UTCTime

The time the most recent patching operation was started on the instance.

instancePatchState_operationEndTime :: Lens' InstancePatchState UTCTime

The time the most recent patching operation completed on the instance.

instancePatchState_operation :: Lens' InstancePatchState PatchOperationType

The type of patching operation that was performed: or

SCAN assesses the patch compliance state.

INSTALL installs missing patches.