Amazonka.SSM.Types.InventoryItemSchema

Description

Documentation

data InventoryItemSchema

The inventory item schema definition. Users can use this to compose inventory query filters.

See: newInventoryItemSchema smart constructor.

Constructors

InventoryItemSchema'

Fields

version :: Maybe Text

The schema version for the inventory item.

displayName :: Maybe Text

The alias name of the inventory type. The alias name is used for display purposes.

typeName :: Text

The name of the inventory type. Default inventory item type names start with Amazon Web Services. Custom inventory type names will start with Custom. Default inventory item types include the following: AWS:AWSComponent, AWS:Application, AWS:InstanceInformation, AWS:Network, and AWS:WindowsUpdate.

attributes :: NonEmpty InventoryItemAttribute

The schema attributes for inventory. This contains data type and attribute name.

newInventoryItemSchema

Arguments

:: Text
$sel:typeName:InventoryItemSchema'

-> NonEmpty InventoryItemAttribute
$sel:attributes:InventoryItemSchema'

-> InventoryItemSchema

Create a value of InventoryItemSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:InventoryItemSchema', inventoryItemSchema_version - The schema version for the inventory item.

$sel:displayName:InventoryItemSchema', inventoryItemSchema_displayName - The alias name of the inventory type. The alias name is used for display purposes.

$sel:typeName:InventoryItemSchema', inventoryItemSchema_typeName - The name of the inventory type. Default inventory item type names start with Amazon Web Services. Custom inventory type names will start with Custom. Default inventory item types include the following: AWS:AWSComponent, AWS:Application, AWS:InstanceInformation, AWS:Network, and AWS:WindowsUpdate.

$sel:attributes:InventoryItemSchema', inventoryItemSchema_attributes - The schema attributes for inventory. This contains data type and attribute name.

inventoryItemSchema_version :: Lens' InventoryItemSchema (Maybe Text)

The schema version for the inventory item.

inventoryItemSchema_displayName :: Lens' InventoryItemSchema (Maybe Text)

The alias name of the inventory type. The alias name is used for display purposes.

inventoryItemSchema_typeName :: Lens' InventoryItemSchema Text

The name of the inventory type. Default inventory item type names start with Amazon Web Services. Custom inventory type names will start with Custom. Default inventory item types include the following: AWS:AWSComponent, AWS:Application, AWS:InstanceInformation, AWS:Network, and AWS:WindowsUpdate.

inventoryItemSchema_attributes :: Lens' InventoryItemSchema (NonEmpty InventoryItemAttribute)

The schema attributes for inventory. This contains data type and attribute name.