Amazonka.SSM.Types.MaintenanceWindowIdentity

Description

Documentation

data MaintenanceWindowIdentity

Information about the maintenance window.

See: newMaintenanceWindowIdentity smart constructor.

Constructors

MaintenanceWindowIdentity'

Fields

enabled :: Maybe Bool

Indicates whether the maintenance window is enabled.

schedule :: Maybe Text

The schedule of the maintenance window in the form of a cron or rate expression.

nextExecutionTime :: Maybe Text

The next time the maintenance window will actually run, taking into account any specified times for the maintenance window to become active or inactive.

scheduleOffset :: Maybe Natural

The number of days to wait to run a maintenance window after the scheduled cron expression date and time.

endDate :: Maybe Text

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become inactive.

scheduleTimezone :: Maybe Text

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format.

startDate :: Maybe Text

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become active.

name :: Maybe Text

The name of the maintenance window.

cutoff :: Maybe Natural

The number of hours before the end of the maintenance window that Amazon Web Services Systems Manager stops scheduling new tasks for execution.

description :: Maybe (Sensitive Text)

A description of the maintenance window.

duration :: Maybe Natural

The duration of the maintenance window in hours.

windowId :: Maybe Text

The ID of the maintenance window.

newMaintenanceWindowIdentity :: MaintenanceWindowIdentity

Create a value of MaintenanceWindowIdentity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:MaintenanceWindowIdentity', maintenanceWindowIdentity_enabled - Indicates whether the maintenance window is enabled.

$sel:schedule:MaintenanceWindowIdentity', maintenanceWindowIdentity_schedule - The schedule of the maintenance window in the form of a cron or rate expression.

$sel:nextExecutionTime:MaintenanceWindowIdentity', maintenanceWindowIdentity_nextExecutionTime - The next time the maintenance window will actually run, taking into account any specified times for the maintenance window to become active or inactive.

$sel:scheduleOffset:MaintenanceWindowIdentity', maintenanceWindowIdentity_scheduleOffset - The number of days to wait to run a maintenance window after the scheduled cron expression date and time.

$sel:endDate:MaintenanceWindowIdentity', maintenanceWindowIdentity_endDate - The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become inactive.

$sel:scheduleTimezone:MaintenanceWindowIdentity', maintenanceWindowIdentity_scheduleTimezone - The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format.

$sel:startDate:MaintenanceWindowIdentity', maintenanceWindowIdentity_startDate - The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become active.

$sel:name:MaintenanceWindowIdentity', maintenanceWindowIdentity_name - The name of the maintenance window.

$sel:cutoff:MaintenanceWindowIdentity', maintenanceWindowIdentity_cutoff - The number of hours before the end of the maintenance window that Amazon Web Services Systems Manager stops scheduling new tasks for execution.

$sel:description:MaintenanceWindowIdentity', maintenanceWindowIdentity_description - A description of the maintenance window.

$sel:duration:MaintenanceWindowIdentity', maintenanceWindowIdentity_duration - The duration of the maintenance window in hours.

$sel:windowId:MaintenanceWindowIdentity', maintenanceWindowIdentity_windowId - The ID of the maintenance window.

maintenanceWindowIdentity_enabled :: Lens' MaintenanceWindowIdentity (Maybe Bool)

Indicates whether the maintenance window is enabled.

maintenanceWindowIdentity_schedule :: Lens' MaintenanceWindowIdentity (Maybe Text)

The schedule of the maintenance window in the form of a cron or rate expression.

maintenanceWindowIdentity_nextExecutionTime :: Lens' MaintenanceWindowIdentity (Maybe Text)

The next time the maintenance window will actually run, taking into account any specified times for the maintenance window to become active or inactive.

maintenanceWindowIdentity_scheduleOffset :: Lens' MaintenanceWindowIdentity (Maybe Natural)

The number of days to wait to run a maintenance window after the scheduled cron expression date and time.

maintenanceWindowIdentity_endDate :: Lens' MaintenanceWindowIdentity (Maybe Text)

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become inactive.

maintenanceWindowIdentity_scheduleTimezone :: Lens' MaintenanceWindowIdentity (Maybe Text)

The time zone that the scheduled maintenance window executions are based on, in Internet Assigned Numbers Authority (IANA) format.

maintenanceWindowIdentity_startDate :: Lens' MaintenanceWindowIdentity (Maybe Text)

The date and time, in ISO-8601 Extended format, for when the maintenance window is scheduled to become active.

maintenanceWindowIdentity_name :: Lens' MaintenanceWindowIdentity (Maybe Text)

The name of the maintenance window.

maintenanceWindowIdentity_cutoff :: Lens' MaintenanceWindowIdentity (Maybe Natural)

The number of hours before the end of the maintenance window that Amazon Web Services Systems Manager stops scheduling new tasks for execution.

maintenanceWindowIdentity_description :: Lens' MaintenanceWindowIdentity (Maybe Text)

A description of the maintenance window.

maintenanceWindowIdentity_duration :: Lens' MaintenanceWindowIdentity (Maybe Natural)

The duration of the maintenance window in hours.

maintenanceWindowIdentity_windowId :: Lens' MaintenanceWindowIdentity (Maybe Text)

The ID of the maintenance window.