Amazonka.SSM.Types.MaintenanceWindowStepFunctionsParameters

Description

Documentation

data MaintenanceWindowStepFunctionsParameters

The parameters for a STEP_FUNCTIONS task.

For information about specifying and updating task parameters, see RegisterTaskWithMaintenanceWindow and UpdateMaintenanceWindowTask.

LoggingInfo has been deprecated. To specify an Amazon Simple Storage Service (Amazon S3) bucket to contain logs, instead use the OutputS3BucketName and OutputS3KeyPrefix options in the TaskInvocationParameters structure. For information about how Amazon Web Services Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

TaskParameters has been deprecated. To specify parameters to pass to a task when it runs, instead use the Parameters option in the TaskInvocationParameters structure. For information about how Systems Manager handles these options for the supported maintenance window task types, see MaintenanceWindowTaskInvocationParameters.

For Step Functions tasks, Systems Manager ignores any values specified for TaskParameters and LoggingInfo.

See: newMaintenanceWindowStepFunctionsParameters smart constructor.

Constructors

MaintenanceWindowStepFunctionsParameters'

Fields

input :: Maybe (Sensitive Text)

The inputs for the STEP_FUNCTIONS task.

name :: Maybe Text

The name of the STEP_FUNCTIONS task.

newMaintenanceWindowStepFunctionsParameters :: MaintenanceWindowStepFunctionsParameters

Create a value of MaintenanceWindowStepFunctionsParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:input:MaintenanceWindowStepFunctionsParameters', maintenanceWindowStepFunctionsParameters_input - The inputs for the STEP_FUNCTIONS task.

$sel:name:MaintenanceWindowStepFunctionsParameters', maintenanceWindowStepFunctionsParameters_name - The name of the STEP_FUNCTIONS task.

maintenanceWindowStepFunctionsParameters_input :: Lens' MaintenanceWindowStepFunctionsParameters (Maybe Text)

The inputs for the STEP_FUNCTIONS task.

maintenanceWindowStepFunctionsParameters_name :: Lens' MaintenanceWindowStepFunctionsParameters (Maybe Text)

The name of the STEP_FUNCTIONS task.