Amazonka.SSM.UnlabelParameterVersion

Description

Remove a label or labels from a parameter.

Creating a Request

data UnlabelParameterVersion

See: newUnlabelParameterVersion smart constructor.

Constructors

UnlabelParameterVersion'

Fields

name :: Text

The name of the parameter from which you want to delete one or more labels.

parameterVersion :: Integer

The specific version of the parameter which you want to delete one or more labels from. If it isn't present, the call will fail.

labels :: NonEmpty Text

One or more labels to delete from the specified parameter version.

newUnlabelParameterVersion

Arguments

:: Text
$sel:name:UnlabelParameterVersion'

-> Integer
$sel:parameterVersion:UnlabelParameterVersion'

-> NonEmpty Text
$sel:labels:UnlabelParameterVersion'

-> UnlabelParameterVersion

Create a value of UnlabelParameterVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UnlabelParameterVersion', unlabelParameterVersion_name - The name of the parameter from which you want to delete one or more labels.

$sel:parameterVersion:UnlabelParameterVersion', unlabelParameterVersion_parameterVersion - The specific version of the parameter which you want to delete one or more labels from. If it isn't present, the call will fail.

$sel:labels:UnlabelParameterVersion', unlabelParameterVersion_labels - One or more labels to delete from the specified parameter version.

Request Lenses

unlabelParameterVersion_name :: Lens' UnlabelParameterVersion Text

The name of the parameter from which you want to delete one or more labels.

unlabelParameterVersion_parameterVersion :: Lens' UnlabelParameterVersion Integer

The specific version of the parameter which you want to delete one or more labels from. If it isn't present, the call will fail.

unlabelParameterVersion_labels :: Lens' UnlabelParameterVersion (NonEmpty Text)

One or more labels to delete from the specified parameter version.

Destructuring the Response

data UnlabelParameterVersionResponse

See: newUnlabelParameterVersionResponse smart constructor.

Constructors

UnlabelParameterVersionResponse'

Fields

invalidLabels :: Maybe (NonEmpty Text)

The labels that aren't attached to the given parameter version.

removedLabels :: Maybe (NonEmpty Text)

A list of all labels deleted from the parameter.

httpStatus :: Int

The response's http status code.

newUnlabelParameterVersionResponse

Arguments

:: Int
$sel:httpStatus:UnlabelParameterVersionResponse'

-> UnlabelParameterVersionResponse

Create a value of UnlabelParameterVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invalidLabels:UnlabelParameterVersionResponse', unlabelParameterVersionResponse_invalidLabels - The labels that aren't attached to the given parameter version.

$sel:removedLabels:UnlabelParameterVersionResponse', unlabelParameterVersionResponse_removedLabels - A list of all labels deleted from the parameter.

$sel:httpStatus:UnlabelParameterVersionResponse', unlabelParameterVersionResponse_httpStatus - The response's http status code.

Response Lenses

unlabelParameterVersionResponse_invalidLabels :: Lens' UnlabelParameterVersionResponse (Maybe (NonEmpty Text))

The labels that aren't attached to the given parameter version.

unlabelParameterVersionResponse_removedLabels :: Lens' UnlabelParameterVersionResponse (Maybe (NonEmpty Text))

A list of all labels deleted from the parameter.

unlabelParameterVersionResponse_httpStatus :: Lens' UnlabelParameterVersionResponse Int

The response's http status code.