Amazonka.SSM.UpdateServiceSetting

Description

ServiceSetting is an account-level setting for an Amazon Web Services service. This setting defines how a user interacts with or uses a service or a feature of a service. For example, if an Amazon Web Services service charges money to the account based on feature or service usage, then the Amazon Web Services service team might create a default setting of "false". This means the user can't use this feature unless they change the setting to "true" and intentionally opt in for a paid feature.

Services map a SettingId object to a setting value. Amazon Web Services services teams define the default value for a SettingId. You can't create a new SettingId, but you can overwrite the default value if you have the ssm:UpdateServiceSetting permission for the setting. Use the GetServiceSetting API operation to view the current value. Or, use the ResetServiceSetting to change the value back to the original value defined by the Amazon Web Services service team.

Update the service setting for the account.

Creating a Request

data UpdateServiceSetting

The request body of the UpdateServiceSetting API operation.

See: newUpdateServiceSetting smart constructor.

Constructors

UpdateServiceSetting'

Fields

settingId :: Text

The Amazon Resource Name (ARN) of the service setting to reset. For example, arn:aws:ssm:us-east-1:111122223333:servicesetting/ssm/parameter-store/high-throughput-enabled. The setting ID can be one of the following.

/ssm/automation/customer-script-log-destination

/ssm/automation/customer-script-log-group-name

/ssm/documents/console/public-sharing-permission

/ssm/parameter-store/default-parameter-tier

/ssm/parameter-store/high-throughput-enabled

/ssm/managed-instance/activation-tier

settingValue :: Text

The new value to specify for the service setting. The following list specifies the available values for each setting.

/ssm/parameter-store/default-parameter-tier: Standard, Advanced, Intelligent-Tiering

/ssm/parameter-store/high-throughput-enabled: true or false

/ssm/managed-instance/activation-tier: true or false

/ssm/automation/customer-script-log-destination: CloudWatch

/ssm/automation/customer-script-log-group-name: the name of an Amazon CloudWatch Logs log group

/ssm/documents/console/public-sharing-permission: Enable or Disable

/ssm/managed-instance/activation-tier: standard or advanced

newUpdateServiceSetting

Arguments

:: Text
$sel:settingId:UpdateServiceSetting'

-> Text
$sel:settingValue:UpdateServiceSetting'

-> UpdateServiceSetting

Create a value of UpdateServiceSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:settingId:UpdateServiceSetting', updateServiceSetting_settingId - The Amazon Resource Name (ARN) of the service setting to reset. For example, arn:aws:ssm:us-east-1:111122223333:servicesetting/ssm/parameter-store/high-throughput-enabled. The setting ID can be one of the following.

/ssm/automation/customer-script-log-destination

/ssm/automation/customer-script-log-group-name

/ssm/documents/console/public-sharing-permission

/ssm/parameter-store/default-parameter-tier

/ssm/parameter-store/high-throughput-enabled

/ssm/managed-instance/activation-tier

$sel:settingValue:UpdateServiceSetting', updateServiceSetting_settingValue - The new value to specify for the service setting. The following list specifies the available values for each setting.

/ssm/parameter-store/default-parameter-tier: Standard, Advanced, Intelligent-Tiering

/ssm/parameter-store/high-throughput-enabled: true or false

/ssm/managed-instance/activation-tier: true or false

/ssm/automation/customer-script-log-destination: CloudWatch

/ssm/automation/customer-script-log-group-name: the name of an Amazon CloudWatch Logs log group

/ssm/documents/console/public-sharing-permission: Enable or Disable

/ssm/managed-instance/activation-tier: standard or advanced

Request Lenses

updateServiceSetting_settingId :: Lens' UpdateServiceSetting Text

The Amazon Resource Name (ARN) of the service setting to reset. For example, arn:aws:ssm:us-east-1:111122223333:servicesetting/ssm/parameter-store/high-throughput-enabled. The setting ID can be one of the following.

/ssm/automation/customer-script-log-destination

/ssm/automation/customer-script-log-group-name

/ssm/documents/console/public-sharing-permission

/ssm/parameter-store/default-parameter-tier

/ssm/parameter-store/high-throughput-enabled

/ssm/managed-instance/activation-tier

updateServiceSetting_settingValue :: Lens' UpdateServiceSetting Text

The new value to specify for the service setting. The following list specifies the available values for each setting.

/ssm/parameter-store/default-parameter-tier: Standard, Advanced, Intelligent-Tiering

/ssm/parameter-store/high-throughput-enabled: true or false

/ssm/managed-instance/activation-tier: true or false

/ssm/automation/customer-script-log-destination: CloudWatch

/ssm/automation/customer-script-log-group-name: the name of an Amazon CloudWatch Logs log group

/ssm/documents/console/public-sharing-permission: Enable or Disable

/ssm/managed-instance/activation-tier: standard or advanced

Destructuring the Response

data UpdateServiceSettingResponse

The result body of the UpdateServiceSetting API operation.

See: newUpdateServiceSettingResponse smart constructor.

Constructors

UpdateServiceSettingResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateServiceSettingResponse

Arguments

:: Int
$sel:httpStatus:UpdateServiceSettingResponse'

-> UpdateServiceSettingResponse

Create a value of UpdateServiceSettingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateServiceSettingResponse', updateServiceSettingResponse_httpStatus - The response's http status code.

Response Lenses

updateServiceSettingResponse_httpStatus :: Lens' UpdateServiceSettingResponse Int

The response's http status code.