Amazonka.SSMContacts.CreateContact

Description

Contacts are either the contacts that Incident Manager engages during an incident or the escalation plans that Incident Manager uses to engage contacts in phases during an incident.

Creating a Request

data CreateContact

See: newCreateContact smart constructor.

Constructors

CreateContact'

Fields

idempotencyToken :: Maybe Text

A token ensuring that the operation is called only once with the specified details.

displayName :: Maybe Text

The full name of the contact or escalation plan.

tags :: Maybe [Tag]

Adds a tag to the target. You can only tag resources created in the first Region of your replication set.

alias :: Text

The short name to quickly identify a contact or escalation plan. The contact alias must be unique and identifiable.

type' :: ContactType

To create an escalation plan use ESCALATION. To create a contact use PERSONAL.

plan :: Plan

A list of stages. A contact has an engagement plan with stages that contact specified contact channels. An escalation plan uses stages that contact specified contacts.

newCreateContact

Arguments

:: Text
$sel:alias:CreateContact'

-> ContactType
$sel:type':CreateContact'

-> Plan
$sel:plan:CreateContact'

-> CreateContact

Create a value of CreateContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idempotencyToken:CreateContact', createContact_idempotencyToken - A token ensuring that the operation is called only once with the specified details.

$sel:displayName:CreateContact', createContact_displayName - The full name of the contact or escalation plan.

$sel:tags:CreateContact', createContact_tags - Adds a tag to the target. You can only tag resources created in the first Region of your replication set.

$sel:alias:CreateContact', createContact_alias - The short name to quickly identify a contact or escalation plan. The contact alias must be unique and identifiable.

$sel:type':CreateContact', createContact_type - To create an escalation plan use ESCALATION. To create a contact use PERSONAL.

$sel:plan:CreateContact', createContact_plan - A list of stages. A contact has an engagement plan with stages that contact specified contact channels. An escalation plan uses stages that contact specified contacts.

Request Lenses

createContact_idempotencyToken :: Lens' CreateContact (Maybe Text)

A token ensuring that the operation is called only once with the specified details.

createContact_displayName :: Lens' CreateContact (Maybe Text)

The full name of the contact or escalation plan.

createContact_tags :: Lens' CreateContact (Maybe [Tag])

Adds a tag to the target. You can only tag resources created in the first Region of your replication set.

createContact_alias :: Lens' CreateContact Text

The short name to quickly identify a contact or escalation plan. The contact alias must be unique and identifiable.

createContact_type :: Lens' CreateContact ContactType

To create an escalation plan use ESCALATION. To create a contact use PERSONAL.

createContact_plan :: Lens' CreateContact Plan

A list of stages. A contact has an engagement plan with stages that contact specified contact channels. An escalation plan uses stages that contact specified contacts.

Destructuring the Response

data CreateContactResponse

See: newCreateContactResponse smart constructor.

Constructors

CreateContactResponse'

Fields

httpStatus :: Int

The response's http status code.

contactArn :: Text

The Amazon Resource Name (ARN) of the created contact or escalation plan.

newCreateContactResponse

Arguments

:: Int
$sel:httpStatus:CreateContactResponse'

-> Text
$sel:contactArn:CreateContactResponse'

-> CreateContactResponse

Create a value of CreateContactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateContactResponse', createContactResponse_httpStatus - The response's http status code.

$sel:contactArn:CreateContactResponse', createContactResponse_contactArn - The Amazon Resource Name (ARN) of the created contact or escalation plan.

Response Lenses

createContactResponse_httpStatus :: Lens' CreateContactResponse Int

The response's http status code.

createContactResponse_contactArn :: Lens' CreateContactResponse Text

The Amazon Resource Name (ARN) of the created contact or escalation plan.