Amazonka.SSMContacts.CreateContactChannel

Description

A contact channel is the method that Incident Manager uses to engage your contact.

Creating a Request

data CreateContactChannel

See: newCreateContactChannel smart constructor.

Constructors

CreateContactChannel'

Fields

idempotencyToken :: Maybe Text

A token ensuring that the operation is called only once with the specified details.

deferActivation :: Maybe Bool

If you want to activate the channel at a later time, you can choose to defer activation. Incident Manager can't engage your contact channel until it has been activated.

contactId :: Text

The Amazon Resource Name (ARN) of the contact you are adding the contact channel to.

name :: Text

The name of the contact channel.

type' :: ChannelType

Incident Manager supports three types of contact channels:

SMS

VOICE

EMAIL

deliveryAddress :: ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel. The format is dependent on the type of the contact channel. The following are the expected formats:

SMS - '+' followed by the country code and phone number

VOICE - '+' followed by the country code and phone number

EMAIL - any standard email format

newCreateContactChannel

Arguments

:: Text
$sel:contactId:CreateContactChannel'

-> Text
$sel:name:CreateContactChannel'

-> ChannelType
$sel:type':CreateContactChannel'

-> ContactChannelAddress
$sel:deliveryAddress:CreateContactChannel'

-> CreateContactChannel

Create a value of CreateContactChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idempotencyToken:CreateContactChannel', createContactChannel_idempotencyToken - A token ensuring that the operation is called only once with the specified details.

$sel:deferActivation:CreateContactChannel', createContactChannel_deferActivation - If you want to activate the channel at a later time, you can choose to defer activation. Incident Manager can't engage your contact channel until it has been activated.

$sel:contactId:CreateContactChannel', createContactChannel_contactId - The Amazon Resource Name (ARN) of the contact you are adding the contact channel to.

$sel:name:CreateContactChannel', createContactChannel_name - The name of the contact channel.

$sel:type':CreateContactChannel', createContactChannel_type - Incident Manager supports three types of contact channels:

SMS

VOICE

EMAIL

$sel:deliveryAddress:CreateContactChannel', createContactChannel_deliveryAddress - The details that Incident Manager uses when trying to engage the contact channel. The format is dependent on the type of the contact channel. The following are the expected formats:

SMS - '+' followed by the country code and phone number

VOICE - '+' followed by the country code and phone number

EMAIL - any standard email format

Request Lenses

createContactChannel_idempotencyToken :: Lens' CreateContactChannel (Maybe Text)

A token ensuring that the operation is called only once with the specified details.

createContactChannel_deferActivation :: Lens' CreateContactChannel (Maybe Bool)

If you want to activate the channel at a later time, you can choose to defer activation. Incident Manager can't engage your contact channel until it has been activated.

createContactChannel_contactId :: Lens' CreateContactChannel Text

The Amazon Resource Name (ARN) of the contact you are adding the contact channel to.

createContactChannel_name :: Lens' CreateContactChannel Text

The name of the contact channel.

createContactChannel_type :: Lens' CreateContactChannel ChannelType

Incident Manager supports three types of contact channels:

SMS

VOICE

EMAIL

createContactChannel_deliveryAddress :: Lens' CreateContactChannel ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel. The format is dependent on the type of the contact channel. The following are the expected formats:

SMS - '+' followed by the country code and phone number

VOICE - '+' followed by the country code and phone number

EMAIL - any standard email format

Destructuring the Response

data CreateContactChannelResponse

See: newCreateContactChannelResponse smart constructor.

Constructors

CreateContactChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

contactChannelArn :: Text

The Amazon Resource Name (ARN) of the contact channel.

newCreateContactChannelResponse

Arguments

:: Int
$sel:httpStatus:CreateContactChannelResponse'

-> Text
$sel:contactChannelArn:CreateContactChannelResponse'

-> CreateContactChannelResponse

Create a value of CreateContactChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateContactChannelResponse', createContactChannelResponse_httpStatus - The response's http status code.

$sel:contactChannelArn:CreateContactChannelResponse', createContactChannelResponse_contactChannelArn - The Amazon Resource Name (ARN) of the contact channel.

Response Lenses

createContactChannelResponse_httpStatus :: Lens' CreateContactChannelResponse Int

The response's http status code.

createContactChannelResponse_contactChannelArn :: Lens' CreateContactChannelResponse Text

The Amazon Resource Name (ARN) of the contact channel.