Amazonka.SSMContacts.DeleteContactChannel

Description

To no longer receive engagements on a contact channel, you can delete the channel from a contact. Deleting the contact channel removes it from the contact's engagement plan. If you delete the only contact channel for a contact, you won't be able to engage that contact during an incident.

Creating a Request

data DeleteContactChannel

See: newDeleteContactChannel smart constructor.

Constructors

DeleteContactChannel'

Fields

contactChannelId :: Text

The Amazon Resource Name (ARN) of the contact channel.

newDeleteContactChannel

Arguments

:: Text
$sel:contactChannelId:DeleteContactChannel'

-> DeleteContactChannel

Create a value of DeleteContactChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactChannelId:DeleteContactChannel', deleteContactChannel_contactChannelId - The Amazon Resource Name (ARN) of the contact channel.

Request Lenses

deleteContactChannel_contactChannelId :: Lens' DeleteContactChannel Text

The Amazon Resource Name (ARN) of the contact channel.

Destructuring the Response

data DeleteContactChannelResponse

See: newDeleteContactChannelResponse smart constructor.

Constructors

DeleteContactChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteContactChannelResponse

Arguments

:: Int
$sel:httpStatus:DeleteContactChannelResponse'

-> DeleteContactChannelResponse

Create a value of DeleteContactChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteContactChannelResponse', deleteContactChannelResponse_httpStatus - The response's http status code.

Response Lenses

deleteContactChannelResponse_httpStatus :: Lens' DeleteContactChannelResponse Int

The response's http status code.