Amazonka.SSMContacts.Lens

Description

Operations

ListPagesByEngagement

listPagesByEngagement_nextToken :: Lens' ListPagesByEngagement (Maybe Text)

The pagination token to continue to the next page of results.

listPagesByEngagement_maxResults :: Lens' ListPagesByEngagement (Maybe Natural)

The maximum number of engagements to contact channels to list per page of results.

listPagesByEngagement_engagementId :: Lens' ListPagesByEngagement Text

The Amazon Resource Name (ARN) of the engagement.

listPagesByEngagementResponse_nextToken :: Lens' ListPagesByEngagementResponse (Maybe Text)

The pagination token to continue to the next page of results.

listPagesByEngagementResponse_httpStatus :: Lens' ListPagesByEngagementResponse Int

The response's http status code.

listPagesByEngagementResponse_pages :: Lens' ListPagesByEngagementResponse [Page]

The list of engagements to contact channels.

ListEngagements

listEngagements_timeRangeValue :: Lens' ListEngagements (Maybe TimeRange)

The time range to lists engagements for an incident.

listEngagements_nextToken :: Lens' ListEngagements (Maybe Text)

The pagination token to continue to the next page of results.

listEngagements_incidentId :: Lens' ListEngagements (Maybe Text)

The Amazon Resource Name (ARN) of the incident you're listing engagements for.

listEngagements_maxResults :: Lens' ListEngagements (Maybe Natural)

The maximum number of engagements per page of results.

listEngagementsResponse_nextToken :: Lens' ListEngagementsResponse (Maybe Text)

The pagination token to continue to the next page of results.

listEngagementsResponse_httpStatus :: Lens' ListEngagementsResponse Int

The response's http status code.

listEngagementsResponse_engagements :: Lens' ListEngagementsResponse [Engagement]

A list of each engagement that occurred during the specified time range of an incident.

ListContactChannels

listContactChannels_nextToken :: Lens' ListContactChannels (Maybe Text)

The pagination token to continue to the next page of results.

listContactChannels_maxResults :: Lens' ListContactChannels (Maybe Natural)

The maximum number of contact channels per page.

listContactChannels_contactId :: Lens' ListContactChannels Text

The Amazon Resource Name (ARN) of the contact.

listContactChannelsResponse_nextToken :: Lens' ListContactChannelsResponse (Maybe Text)

The pagination token to continue to the next page of results.

listContactChannelsResponse_httpStatus :: Lens' ListContactChannelsResponse Int

The response's http status code.

listContactChannelsResponse_contactChannels :: Lens' ListContactChannelsResponse [ContactChannel]

A list of contact channels related to the specified contact.

ActivateContactChannel

activateContactChannel_contactChannelId :: Lens' ActivateContactChannel Text

The Amazon Resource Name (ARN) of the contact channel.

activateContactChannel_activationCode :: Lens' ActivateContactChannel Text

The code sent to the contact channel when it was created in the contact.

activateContactChannelResponse_httpStatus :: Lens' ActivateContactChannelResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceARN :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the contact or escalation plan.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

The tags related to the contact or escalation plan.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

StartEngagement

startEngagement_idempotencyToken :: Lens' StartEngagement (Maybe Text)

A token ensuring that the operation is called only once with the specified details.

startEngagement_publicSubject :: Lens' StartEngagement (Maybe Text)

The insecure subject of the message that was sent to the contact. Use this field for engagements to SMS.

startEngagement_publicContent :: Lens' StartEngagement (Maybe Text)

The insecure content of the message that was sent to the contact. Use this field for engagements to SMS.

startEngagement_incidentId :: Lens' StartEngagement (Maybe Text)

The ARN of the incident that the engagement is part of.

startEngagement_contactId :: Lens' StartEngagement Text

The Amazon Resource Name (ARN) of the contact being engaged.

startEngagement_sender :: Lens' StartEngagement Text

The user that started the engagement.

startEngagement_subject :: Lens' StartEngagement Text

The secure subject of the message that was sent to the contact. Use this field for engagements to VOICE or EMAIL.

startEngagement_content :: Lens' StartEngagement Text

The secure content of the message that was sent to the contact. Use this field for engagements to VOICE or EMAIL.

startEngagementResponse_httpStatus :: Lens' StartEngagementResponse Int

The response's http status code.

startEngagementResponse_engagementArn :: Lens' StartEngagementResponse Text

The ARN of the engagement.

DeactivateContactChannel

deactivateContactChannel_contactChannelId :: Lens' DeactivateContactChannel Text

The Amazon Resource Name (ARN) of the contact channel you're deactivating.

deactivateContactChannelResponse_httpStatus :: Lens' DeactivateContactChannelResponse Int

The response's http status code.

AcceptPage

acceptPage_note :: Lens' AcceptPage (Maybe Text)

Information provided by the user when the user acknowledges the page.

acceptPage_contactChannelId :: Lens' AcceptPage (Maybe Text)

The ARN of the contact channel.

acceptPage_acceptCodeValidation :: Lens' AcceptPage (Maybe AcceptCodeValidation)

An optional field that Incident Manager uses to ENFORCE AcceptCode validation when acknowledging an page. Acknowledgement can occur by replying to a page, or when entering the AcceptCode in the console. Enforcing AcceptCode validation causes Incident Manager to verify that the code entered by the user matches the code sent by Incident Manager with the page.

Incident Manager can also IGNORE AcceptCode validation. Ignoring AcceptCode validation causes Incident Manager to accept any value entered for the AcceptCode.

acceptPage_pageId :: Lens' AcceptPage Text

The Amazon Resource Name (ARN) of the engagement to a contact channel.

acceptPage_acceptType :: Lens' AcceptPage AcceptType

The type indicates if the page was DELIVERED or READ.

acceptPage_acceptCode :: Lens' AcceptPage Text

The accept code is a 6-digit code used to acknowledge the page.

acceptPageResponse_httpStatus :: Lens' AcceptPageResponse Int

The response's http status code.

ListPageReceipts

listPageReceipts_nextToken :: Lens' ListPageReceipts (Maybe Text)

The pagination token to continue to the next page of results.

listPageReceipts_maxResults :: Lens' ListPageReceipts (Maybe Natural)

The maximum number of acknowledgements per page of results.

listPageReceipts_pageId :: Lens' ListPageReceipts Text

The Amazon Resource Name (ARN) of the engagement to a specific contact channel.

listPageReceiptsResponse_nextToken :: Lens' ListPageReceiptsResponse (Maybe Text)

The pagination token to continue to the next page of results.

listPageReceiptsResponse_receipts :: Lens' ListPageReceiptsResponse (Maybe [Receipt])

A list of each acknowledgement.

listPageReceiptsResponse_httpStatus :: Lens' ListPageReceiptsResponse Int

The response's http status code.

GetContact

getContact_contactId :: Lens' GetContact Text

The Amazon Resource Name (ARN) of the contact or escalation plan.

getContactResponse_displayName :: Lens' GetContactResponse (Maybe Text)

The full name of the contact or escalation plan.

getContactResponse_httpStatus :: Lens' GetContactResponse Int

The response's http status code.

getContactResponse_contactArn :: Lens' GetContactResponse Text

The ARN of the contact or escalation plan.

getContactResponse_alias :: Lens' GetContactResponse Text

The alias of the contact or escalation plan. The alias is unique and identifiable.

getContactResponse_type :: Lens' GetContactResponse ContactType

The type of contact, either PERSONAL or ESCALATION.

getContactResponse_plan :: Lens' GetContactResponse Plan

Details about the specific timing or stages and targets of the escalation plan or engagement plan.

DescribePage

describePage_pageId :: Lens' DescribePage Text

The ID of the engagement to a contact channel.

describePageResponse_readTime :: Lens' DescribePageResponse (Maybe UTCTime)

The time that the contact channel acknowledged the engagement.

describePageResponse_publicSubject :: Lens' DescribePageResponse (Maybe Text)

The insecure subject of the message that was sent to the contact. Use this field for engagements to SMS.

describePageResponse_publicContent :: Lens' DescribePageResponse (Maybe Text)

The insecure content of the message that was sent to the contact. Use this field for engagements to SMS.

describePageResponse_deliveryTime :: Lens' DescribePageResponse (Maybe UTCTime)

The time that the contact channel received the engagement.

describePageResponse_incidentId :: Lens' DescribePageResponse (Maybe Text)

The ARN of the incident that engaged the contact channel.

describePageResponse_sentTime :: Lens' DescribePageResponse (Maybe UTCTime)

The time the engagement was sent to the contact channel.

describePageResponse_httpStatus :: Lens' DescribePageResponse Int

The response's http status code.

describePageResponse_pageArn :: Lens' DescribePageResponse Text

The Amazon Resource Name (ARN) of the engagement to a contact channel.

describePageResponse_engagementArn :: Lens' DescribePageResponse Text

The ARN of the engagement that engaged the contact channel.

describePageResponse_contactArn :: Lens' DescribePageResponse Text

The ARN of the contact that was engaged.

describePageResponse_sender :: Lens' DescribePageResponse Text

The user that started the engagement.

describePageResponse_subject :: Lens' DescribePageResponse Text

The secure subject of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

describePageResponse_content :: Lens' DescribePageResponse Text

The secure content of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

DeleteContact

deleteContact_contactId :: Lens' DeleteContact Text

The Amazon Resource Name (ARN) of the contact that you're deleting.

deleteContactResponse_httpStatus :: Lens' DeleteContactResponse Int

The response's http status code.

UpdateContact

updateContact_plan :: Lens' UpdateContact (Maybe Plan)

A list of stages. A contact has an engagement plan with stages for specified contact channels. An escalation plan uses these stages to contact specified contacts.

updateContact_displayName :: Lens' UpdateContact (Maybe Text)

The full name of the contact or escalation plan.

updateContact_contactId :: Lens' UpdateContact Text

The Amazon Resource Name (ARN) of the contact or escalation plan you're updating.

updateContactResponse_httpStatus :: Lens' UpdateContactResponse Int

The response's http status code.

CreateContact

createContact_idempotencyToken :: Lens' CreateContact (Maybe Text)

A token ensuring that the operation is called only once with the specified details.

createContact_displayName :: Lens' CreateContact (Maybe Text)

The full name of the contact or escalation plan.

createContact_tags :: Lens' CreateContact (Maybe [Tag])

Adds a tag to the target. You can only tag resources created in the first Region of your replication set.

createContact_alias :: Lens' CreateContact Text

The short name to quickly identify a contact or escalation plan. The contact alias must be unique and identifiable.

createContact_type :: Lens' CreateContact ContactType

To create an escalation plan use ESCALATION. To create a contact use PERSONAL.

createContact_plan :: Lens' CreateContact Plan

A list of stages. A contact has an engagement plan with stages that contact specified contact channels. An escalation plan uses stages that contact specified contacts.

createContactResponse_httpStatus :: Lens' CreateContactResponse Int

The response's http status code.

createContactResponse_contactArn :: Lens' CreateContactResponse Text

The Amazon Resource Name (ARN) of the created contact or escalation plan.

CreateContactChannel

createContactChannel_idempotencyToken :: Lens' CreateContactChannel (Maybe Text)

A token ensuring that the operation is called only once with the specified details.

createContactChannel_deferActivation :: Lens' CreateContactChannel (Maybe Bool)

If you want to activate the channel at a later time, you can choose to defer activation. Incident Manager can't engage your contact channel until it has been activated.

createContactChannel_contactId :: Lens' CreateContactChannel Text

The Amazon Resource Name (ARN) of the contact you are adding the contact channel to.

createContactChannel_name :: Lens' CreateContactChannel Text

The name of the contact channel.

createContactChannel_type :: Lens' CreateContactChannel ChannelType

Incident Manager supports three types of contact channels:

SMS

VOICE

EMAIL

createContactChannel_deliveryAddress :: Lens' CreateContactChannel ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel. The format is dependent on the type of the contact channel. The following are the expected formats:

SMS - '+' followed by the country code and phone number

VOICE - '+' followed by the country code and phone number

EMAIL - any standard email format

createContactChannelResponse_httpStatus :: Lens' CreateContactChannelResponse Int

The response's http status code.

createContactChannelResponse_contactChannelArn :: Lens' CreateContactChannelResponse Text

The Amazon Resource Name (ARN) of the contact channel.

DeleteContactChannel

deleteContactChannel_contactChannelId :: Lens' DeleteContactChannel Text

The Amazon Resource Name (ARN) of the contact channel.

deleteContactChannelResponse_httpStatus :: Lens' DeleteContactChannelResponse Int

The response's http status code.

UpdateContactChannel

updateContactChannel_name :: Lens' UpdateContactChannel (Maybe Text)

The name of the contact channel.

updateContactChannel_deliveryAddress :: Lens' UpdateContactChannel (Maybe ContactChannelAddress)

The details that Incident Manager uses when trying to engage the contact channel.

updateContactChannel_contactChannelId :: Lens' UpdateContactChannel Text

The Amazon Resource Name (ARN) of the contact channel you want to update.

updateContactChannelResponse_httpStatus :: Lens' UpdateContactChannelResponse Int

The response's http status code.

GetContactChannel

getContactChannel_contactChannelId :: Lens' GetContactChannel Text

The Amazon Resource Name (ARN) of the contact channel you want information about.

getContactChannelResponse_activationStatus :: Lens' GetContactChannelResponse (Maybe ActivationStatus)

A Boolean value indicating if the contact channel has been activated or not.

getContactChannelResponse_httpStatus :: Lens' GetContactChannelResponse Int

The response's http status code.

getContactChannelResponse_contactArn :: Lens' GetContactChannelResponse Text

The ARN of the contact that the channel belongs to.

getContactChannelResponse_contactChannelArn :: Lens' GetContactChannelResponse Text

The ARN of the contact channel.

getContactChannelResponse_name :: Lens' GetContactChannelResponse Text

The name of the contact channel

getContactChannelResponse_type :: Lens' GetContactChannelResponse ChannelType

The type of contact channel. The type is SMS, VOICE, or EMAIL.

getContactChannelResponse_deliveryAddress :: Lens' GetContactChannelResponse ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel.

SendActivationCode

sendActivationCode_contactChannelId :: Lens' SendActivationCode Text

The Amazon Resource Name (ARN) of the contact channel.

sendActivationCodeResponse_httpStatus :: Lens' SendActivationCodeResponse Int

The response's http status code.

StopEngagement

stopEngagement_reason :: Lens' StopEngagement (Maybe Text)

The reason that you're stopping the engagement.

stopEngagement_engagementId :: Lens' StopEngagement Text

The Amazon Resource Name (ARN) of the engagement.

stopEngagementResponse_httpStatus :: Lens' StopEngagementResponse Int

The response's http status code.

DescribeEngagement

describeEngagement_engagementId :: Lens' DescribeEngagement Text

The Amazon Resource Name (ARN) of the engagement you want the details of.

describeEngagementResponse_publicSubject :: Lens' DescribeEngagementResponse (Maybe Text)

The insecure subject of the message that was sent to the contact. Use this field for engagements to SMS.

describeEngagementResponse_startTime :: Lens' DescribeEngagementResponse (Maybe UTCTime)

The time that the engagement started.

describeEngagementResponse_publicContent :: Lens' DescribeEngagementResponse (Maybe Text)

The insecure content of the message that was sent to the contact. Use this field for engagements to SMS.

describeEngagementResponse_stopTime :: Lens' DescribeEngagementResponse (Maybe UTCTime)

The time that the engagement ended.

describeEngagementResponse_incidentId :: Lens' DescribeEngagementResponse (Maybe Text)

The ARN of the incident in which the engagement occurred.

describeEngagementResponse_httpStatus :: Lens' DescribeEngagementResponse Int

The response's http status code.

describeEngagementResponse_contactArn :: Lens' DescribeEngagementResponse Text

The ARN of the escalation plan or contacts involved in the engagement.

describeEngagementResponse_engagementArn :: Lens' DescribeEngagementResponse Text

The ARN of the engagement.

describeEngagementResponse_sender :: Lens' DescribeEngagementResponse Text

The user that started the engagement.

describeEngagementResponse_subject :: Lens' DescribeEngagementResponse Text

The secure subject of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

describeEngagementResponse_content :: Lens' DescribeEngagementResponse Text

The secure content of the message that was sent to the contact. Use this field for engagements to VOICE and EMAIL.

TagResource

tagResource_resourceARN :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the contact or escalation plan.

tagResource_tags :: Lens' TagResource [Tag]

A list of tags that you are adding to the contact or escalation plan.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceARN :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the contact or escalation plan.

untagResource_tagKeys :: Lens' UntagResource [Text]

The key of the tag that you want to remove.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

GetContactPolicy

getContactPolicy_contactArn :: Lens' GetContactPolicy Text

The Amazon Resource Name (ARN) of the contact or escalation plan.

getContactPolicyResponse_contactArn :: Lens' GetContactPolicyResponse (Maybe Text)

The ARN of the contact or escalation plan.

getContactPolicyResponse_policy :: Lens' GetContactPolicyResponse (Maybe Text)

Details about the resource policy attached to the contact or escalation plan.

getContactPolicyResponse_httpStatus :: Lens' GetContactPolicyResponse Int

The response's http status code.

PutContactPolicy

putContactPolicy_contactArn :: Lens' PutContactPolicy Text

The Amazon Resource Name (ARN) of the contact or escalation plan.

putContactPolicy_policy :: Lens' PutContactPolicy Text

Details of the resource policy.

putContactPolicyResponse_httpStatus :: Lens' PutContactPolicyResponse Int

The response's http status code.

ListContacts

listContacts_aliasPrefix :: Lens' ListContacts (Maybe Text)

Used to list only contacts who's aliases start with the specified prefix.

listContacts_nextToken :: Lens' ListContacts (Maybe Text)

The pagination token to continue to the next page of results.

listContacts_type :: Lens' ListContacts (Maybe ContactType)

The type of contact. A contact is type PERSONAL and an escalation plan is type ESCALATION.

listContacts_maxResults :: Lens' ListContacts (Maybe Natural)

The maximum number of contacts and escalation plans per page of results.

listContactsResponse_nextToken :: Lens' ListContactsResponse (Maybe Text)

The pagination token to continue to the next page of results.

listContactsResponse_contacts :: Lens' ListContactsResponse (Maybe [Contact])

A list of the contacts and escalation plans in your Incident Manager account.

listContactsResponse_httpStatus :: Lens' ListContactsResponse Int

The response's http status code.

ListPagesByContact

listPagesByContact_nextToken :: Lens' ListPagesByContact (Maybe Text)

The pagination token to continue to the next page of results.

listPagesByContact_maxResults :: Lens' ListPagesByContact (Maybe Natural)

The maximum number of engagements to contact channels to list per page of results.

listPagesByContact_contactId :: Lens' ListPagesByContact Text

The Amazon Resource Name (ARN) of the contact you are retrieving engagements for.

listPagesByContactResponse_nextToken :: Lens' ListPagesByContactResponse (Maybe Text)

The pagination token to continue to the next page of results.

listPagesByContactResponse_httpStatus :: Lens' ListPagesByContactResponse Int

The response's http status code.

listPagesByContactResponse_pages :: Lens' ListPagesByContactResponse [Page]

The list of engagements to a contact's contact channel.

Types

ChannelTargetInfo

channelTargetInfo_retryIntervalInMinutes :: Lens' ChannelTargetInfo (Maybe Natural)

The number of minutes to wait to retry sending engagement in the case the engagement initially fails.

channelTargetInfo_contactChannelId :: Lens' ChannelTargetInfo Text

The Amazon Resource Name (ARN) of the contact channel.

Contact

contact_displayName :: Lens' Contact (Maybe Text)

The full name of the contact or escalation plan.

contact_contactArn :: Lens' Contact Text

The Amazon Resource Name (ARN) of the contact or escalation plan.

contact_alias :: Lens' Contact Text

The unique and identifiable alias of the contact or escalation plan.

contact_type :: Lens' Contact ContactType

Refers to the type of contact. A single contact is type PERSONAL and an escalation plan is type ESCALATION.

ContactChannel

contactChannel_type :: Lens' ContactChannel (Maybe ChannelType)

The type of the contact channel. Incident Manager supports three contact methods:

SMS

VOICE

EMAIL

contactChannel_contactChannelArn :: Lens' ContactChannel Text

The Amazon Resource Name (ARN) of the contact channel.

contactChannel_contactArn :: Lens' ContactChannel Text

The ARN of the contact that contains the contact channel.

contactChannel_name :: Lens' ContactChannel Text

The name of the contact channel.

contactChannel_deliveryAddress :: Lens' ContactChannel ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel.

contactChannel_activationStatus :: Lens' ContactChannel ActivationStatus

A Boolean value describing if the contact channel has been activated or not. If the contact channel isn't activated, Incident Manager can't engage the contact through it.

ContactChannelAddress

contactChannelAddress_simpleAddress :: Lens' ContactChannelAddress (Maybe Text)

The format is dependent on the type of the contact channel. The following are the expected formats:

SMS - '+' followed by the country code and phone number

VOICE - '+' followed by the country code and phone number

EMAIL - any standard email format

ContactTargetInfo

contactTargetInfo_contactId :: Lens' ContactTargetInfo (Maybe Text)

The Amazon Resource Name (ARN) of the contact.

contactTargetInfo_isEssential :: Lens' ContactTargetInfo Bool

A Boolean value determining if the contact's acknowledgement stops the progress of stages in the plan.

Engagement

engagement_startTime :: Lens' Engagement (Maybe UTCTime)

The time that the engagement began.

engagement_stopTime :: Lens' Engagement (Maybe UTCTime)

The time that the engagement ended.

engagement_incidentId :: Lens' Engagement (Maybe Text)

The ARN of the incident that's engaging the contact.

engagement_engagementArn :: Lens' Engagement Text

The Amazon Resource Name (ARN) of the engagement.

engagement_contactArn :: Lens' Engagement Text

The ARN of the escalation plan or contact that Incident Manager is engaging.

engagement_sender :: Lens' Engagement Text

The user that started the engagement.

Page

page_readTime :: Lens' Page (Maybe UTCTime)

The time that the contact channel acknowledged engagement.

page_deliveryTime :: Lens' Page (Maybe UTCTime)

The time the message was delivered to the contact channel.

page_incidentId :: Lens' Page (Maybe Text)

The ARN of the incident that's engaging the contact channel.

page_sentTime :: Lens' Page (Maybe UTCTime)

The time that Incident Manager engaged the contact channel.

page_pageArn :: Lens' Page Text

The Amazon Resource Name (ARN) of the page to the contact channel.

page_engagementArn :: Lens' Page Text

The ARN of the engagement that this page is part of.

page_contactArn :: Lens' Page Text

The ARN of the contact that Incident Manager is engaging.

page_sender :: Lens' Page Text

The user that started the engagement.

Plan

plan_stages :: Lens' Plan [Stage]

A list of stages that the escalation plan or engagement plan uses to engage contacts and contact methods.

Receipt

receipt_receiptInfo :: Lens' Receipt (Maybe Text)

Information provided during the page acknowledgement.

receipt_contactChannelArn :: Lens' Receipt (Maybe Text)

The Amazon Resource Name (ARN) of the contact channel Incident Manager engaged.

receipt_receiptType :: Lens' Receipt ReceiptType

The type follows the engagement cycle, SENT, DELIVERED, and READ.

receipt_receiptTime :: Lens' Receipt UTCTime

The time receipt was SENT, DELIVERED, or READ.

Stage

stage_durationInMinutes :: Lens' Stage Natural

The time to wait until beginning the next stage. The duration can only be set to 0 if a target is specified.

stage_targets :: Lens' Stage [Target]

The contacts or contact methods that the escalation plan or engagement plan is engaging.

Tag

tag_value :: Lens' Tag (Maybe Text)

Value of the tag.

tag_key :: Lens' Tag (Maybe Text)

Name of the object key.

Target

target_channelTargetInfo :: Lens' Target (Maybe ChannelTargetInfo)

Information about the contact channel Incident Manager is engaging.

target_contactTargetInfo :: Lens' Target (Maybe ContactTargetInfo)

Information about the contact that Incident Manager is engaging.

TimeRange

timeRange_startTime :: Lens' TimeRange (Maybe UTCTime)

The start of the time range.

timeRange_endTime :: Lens' TimeRange (Maybe UTCTime)

The end of the time range.