Amazonka.SSMContacts.ListContactChannels

Description

Lists all contact channels for the specified contact.

This operation returns paginated results.

Creating a Request

data ListContactChannels

See: newListContactChannels smart constructor.

Constructors

ListContactChannels'

Fields

nextToken :: Maybe Text

The pagination token to continue to the next page of results.

maxResults :: Maybe Natural

The maximum number of contact channels per page.

contactId :: Text

The Amazon Resource Name (ARN) of the contact.

newListContactChannels

Arguments

:: Text
$sel:contactId:ListContactChannels'

-> ListContactChannels

Create a value of ListContactChannels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContactChannels', listContactChannels_nextToken - The pagination token to continue to the next page of results.

$sel:maxResults:ListContactChannels', listContactChannels_maxResults - The maximum number of contact channels per page.

$sel:contactId:ListContactChannels', listContactChannels_contactId - The Amazon Resource Name (ARN) of the contact.

Request Lenses

listContactChannels_nextToken :: Lens' ListContactChannels (Maybe Text)

The pagination token to continue to the next page of results.

listContactChannels_maxResults :: Lens' ListContactChannels (Maybe Natural)

The maximum number of contact channels per page.

listContactChannels_contactId :: Lens' ListContactChannels Text

The Amazon Resource Name (ARN) of the contact.

Destructuring the Response

data ListContactChannelsResponse

See: newListContactChannelsResponse smart constructor.

Constructors

ListContactChannelsResponse'

Fields

nextToken :: Maybe Text

The pagination token to continue to the next page of results.

httpStatus :: Int

The response's http status code.

contactChannels :: [ContactChannel]

A list of contact channels related to the specified contact.

newListContactChannelsResponse

Arguments

:: Int
$sel:httpStatus:ListContactChannelsResponse'

-> ListContactChannelsResponse

Create a value of ListContactChannelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContactChannels', listContactChannelsResponse_nextToken - The pagination token to continue to the next page of results.

$sel:httpStatus:ListContactChannelsResponse', listContactChannelsResponse_httpStatus - The response's http status code.

$sel:contactChannels:ListContactChannelsResponse', listContactChannelsResponse_contactChannels - A list of contact channels related to the specified contact.

Response Lenses

listContactChannelsResponse_nextToken :: Lens' ListContactChannelsResponse (Maybe Text)

The pagination token to continue to the next page of results.

listContactChannelsResponse_httpStatus :: Lens' ListContactChannelsResponse Int

The response's http status code.

listContactChannelsResponse_contactChannels :: Lens' ListContactChannelsResponse [ContactChannel]

A list of contact channels related to the specified contact.