Amazonka.SSMContacts.Types.ContactChannel

Description

Documentation

data ContactChannel

The method that Incident Manager uses to engage a contact.

See: newContactChannel smart constructor.

Constructors

ContactChannel'

Fields

type' :: Maybe ChannelType

The type of the contact channel. Incident Manager supports three contact methods:

SMS

VOICE

EMAIL

contactChannelArn :: Text

The Amazon Resource Name (ARN) of the contact channel.

contactArn :: Text

The ARN of the contact that contains the contact channel.

name :: Text

The name of the contact channel.

deliveryAddress :: ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel.

activationStatus :: ActivationStatus

A Boolean value describing if the contact channel has been activated or not. If the contact channel isn't activated, Incident Manager can't engage the contact through it.

newContactChannel

Arguments

:: Text
$sel:contactChannelArn:ContactChannel'

-> Text
$sel:contactArn:ContactChannel'

-> Text
$sel:name:ContactChannel'

-> ContactChannelAddress
$sel:deliveryAddress:ContactChannel'

-> ActivationStatus
$sel:activationStatus:ContactChannel'

-> ContactChannel

Create a value of ContactChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':ContactChannel', contactChannel_type - The type of the contact channel. Incident Manager supports three contact methods:

SMS

VOICE

EMAIL

$sel:contactChannelArn:ContactChannel', contactChannel_contactChannelArn - The Amazon Resource Name (ARN) of the contact channel.

$sel:contactArn:ContactChannel', contactChannel_contactArn - The ARN of the contact that contains the contact channel.

$sel:name:ContactChannel', contactChannel_name - The name of the contact channel.

$sel:deliveryAddress:ContactChannel', contactChannel_deliveryAddress - The details that Incident Manager uses when trying to engage the contact channel.

$sel:activationStatus:ContactChannel', contactChannel_activationStatus - A Boolean value describing if the contact channel has been activated or not. If the contact channel isn't activated, Incident Manager can't engage the contact through it.

contactChannel_type :: Lens' ContactChannel (Maybe ChannelType)

The type of the contact channel. Incident Manager supports three contact methods:

SMS

VOICE

EMAIL

contactChannel_contactChannelArn :: Lens' ContactChannel Text

The Amazon Resource Name (ARN) of the contact channel.

contactChannel_contactArn :: Lens' ContactChannel Text

The ARN of the contact that contains the contact channel.

contactChannel_name :: Lens' ContactChannel Text

The name of the contact channel.

contactChannel_deliveryAddress :: Lens' ContactChannel ContactChannelAddress

The details that Incident Manager uses when trying to engage the contact channel.

contactChannel_activationStatus :: Lens' ContactChannel ActivationStatus

A Boolean value describing if the contact channel has been activated or not. If the contact channel isn't activated, Incident Manager can't engage the contact through it.