Amazonka.SSMIncidents.CreateResponsePlan

Description

Creates a response plan that automates the initial response to incidents. A response plan engages contacts, starts chat channel collaboration, and initiates runbooks at the beginning of an incident.

Creating a Request

data CreateResponsePlan

See: newCreateResponsePlan smart constructor.

Constructors

CreateResponsePlan'

Fields

clientToken :: Maybe Text

A token ensuring that the action is called only once with the specified details.

actions :: Maybe [Action]

The actions that the response plan starts at the beginning of an incident.

displayName :: Maybe Text

The long format of the response plan name. This field can contain spaces.

chatChannel :: Maybe ChatChannel

The AWS Chatbot chat channel used for collaboration during an incident.

engagements :: Maybe [Text]

The contacts and escalation plans that the response plan engages during an incident.

tags :: Maybe (HashMap Text Text)

A list of tags that you are adding to the response plan.

incidentTemplate :: IncidentTemplate

Details used to create an incident when using this response plan.

name :: Text

The short format name of the response plan. Can't include spaces.

newCreateResponsePlan

Arguments

:: IncidentTemplate
$sel:incidentTemplate:CreateResponsePlan'

-> Text
$sel:name:CreateResponsePlan'

-> CreateResponsePlan

Create a value of CreateResponsePlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateResponsePlan', createResponsePlan_clientToken - A token ensuring that the action is called only once with the specified details.

$sel:actions:CreateResponsePlan', createResponsePlan_actions - The actions that the response plan starts at the beginning of an incident.

$sel:displayName:CreateResponsePlan', createResponsePlan_displayName - The long format of the response plan name. This field can contain spaces.

$sel:chatChannel:CreateResponsePlan', createResponsePlan_chatChannel - The AWS Chatbot chat channel used for collaboration during an incident.

$sel:engagements:CreateResponsePlan', createResponsePlan_engagements - The contacts and escalation plans that the response plan engages during an incident.

$sel:tags:CreateResponsePlan', createResponsePlan_tags - A list of tags that you are adding to the response plan.

$sel:incidentTemplate:CreateResponsePlan', createResponsePlan_incidentTemplate - Details used to create an incident when using this response plan.

$sel:name:CreateResponsePlan', createResponsePlan_name - The short format name of the response plan. Can't include spaces.

Request Lenses

createResponsePlan_clientToken :: Lens' CreateResponsePlan (Maybe Text)

A token ensuring that the action is called only once with the specified details.

createResponsePlan_actions :: Lens' CreateResponsePlan (Maybe [Action])

The actions that the response plan starts at the beginning of an incident.

createResponsePlan_displayName :: Lens' CreateResponsePlan (Maybe Text)

The long format of the response plan name. This field can contain spaces.

createResponsePlan_chatChannel :: Lens' CreateResponsePlan (Maybe ChatChannel)

The AWS Chatbot chat channel used for collaboration during an incident.

createResponsePlan_engagements :: Lens' CreateResponsePlan (Maybe [Text])

The contacts and escalation plans that the response plan engages during an incident.

createResponsePlan_tags :: Lens' CreateResponsePlan (Maybe (HashMap Text Text))

A list of tags that you are adding to the response plan.

createResponsePlan_incidentTemplate :: Lens' CreateResponsePlan IncidentTemplate

Details used to create an incident when using this response plan.

createResponsePlan_name :: Lens' CreateResponsePlan Text

The short format name of the response plan. Can't include spaces.

Destructuring the Response

data CreateResponsePlanResponse

See: newCreateResponsePlanResponse smart constructor.

Constructors

CreateResponsePlanResponse'

Fields

httpStatus :: Int

The response's http status code.

arn :: Text

The Amazon Resource Name (ARN) of the response plan.

newCreateResponsePlanResponse

Arguments

:: Int
$sel:httpStatus:CreateResponsePlanResponse'

-> Text
$sel:arn:CreateResponsePlanResponse'

-> CreateResponsePlanResponse

Create a value of CreateResponsePlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateResponsePlanResponse', createResponsePlanResponse_httpStatus - The response's http status code.

$sel:arn:CreateResponsePlanResponse', createResponsePlanResponse_arn - The Amazon Resource Name (ARN) of the response plan.

Response Lenses

createResponsePlanResponse_httpStatus :: Lens' CreateResponsePlanResponse Int

The response's http status code.

createResponsePlanResponse_arn :: Lens' CreateResponsePlanResponse Text

The Amazon Resource Name (ARN) of the response plan.