Amazonka.SSMIncidents.Types.EventSummary

Description

Documentation

data EventSummary

Details about a timeline event during an incident.

See: newEventSummary smart constructor.

Constructors

EventSummary'

Fields

eventId :: Text

The timeline event ID.

eventTime :: POSIX

The time that the event occurred.

eventType :: Text

The type of event. The timeline event must be Custom Event.

eventUpdatedTime :: POSIX

The time that the timeline event was last updated.

incidentRecordArn :: Text

The Amazon Resource Name (ARN) of the incident that the event happened during.

newEventSummary

Arguments

:: Text
$sel:eventId:EventSummary'

-> UTCTime
$sel:eventTime:EventSummary'

-> Text
$sel:eventType:EventSummary'

-> UTCTime
$sel:eventUpdatedTime:EventSummary'

-> Text
$sel:incidentRecordArn:EventSummary'

-> EventSummary

Create a value of EventSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventId:EventSummary', eventSummary_eventId - The timeline event ID.

$sel:eventTime:EventSummary', eventSummary_eventTime - The time that the event occurred.

$sel:eventType:EventSummary', eventSummary_eventType - The type of event. The timeline event must be Custom Event.

$sel:eventUpdatedTime:EventSummary', eventSummary_eventUpdatedTime - The time that the timeline event was last updated.

$sel:incidentRecordArn:EventSummary', eventSummary_incidentRecordArn - The Amazon Resource Name (ARN) of the incident that the event happened during.

eventSummary_eventId :: Lens' EventSummary Text

The timeline event ID.

eventSummary_eventTime :: Lens' EventSummary UTCTime

The time that the event occurred.

eventSummary_eventType :: Lens' EventSummary Text

The type of event. The timeline event must be Custom Event.

eventSummary_eventUpdatedTime :: Lens' EventSummary UTCTime

The time that the timeline event was last updated.

eventSummary_incidentRecordArn :: Lens' EventSummary Text

The Amazon Resource Name (ARN) of the incident that the event happened during.