Amazonka.SSMIncidents.Types.IncidentRecord

Description

Documentation

data IncidentRecord

The record of the incident that's created when an incident occurs.

See: newIncidentRecord smart constructor.

Constructors

IncidentRecord'

Fields

summary :: Maybe Text

The summary of the incident. The summary is a brief synopsis of what occurred, what is currently happening, and context.

notificationTargets :: Maybe [NotificationTargetItem]

The SNS targets that are notified when updates are made to an incident.

resolvedTime :: Maybe POSIX

The time at which the incident was resolved. This appears as a timeline event.

chatChannel :: Maybe ChatChannel

The chat channel used for collaboration during an incident.

automationExecutions :: Maybe [AutomationExecution]

The runbook, or automation document, that's run at the beginning of the incident.

arn :: Text

The Amazon Resource Name (ARN) of the incident record.

creationTime :: POSIX

The time that Incident Manager created the incident record.

dedupeString :: Text

The string Incident Manager uses to prevent duplicate incidents from being created by the same incident.

impact :: Natural

The impact of the incident on customers and applications.

incidentRecordSource :: IncidentRecordSource

Details about the action that started the incident.

lastModifiedBy :: Text

Who modified the incident most recently.

lastModifiedTime :: POSIX

The time at which the incident was most recently modified.

status :: IncidentRecordStatus

The current status of the incident.

title :: Text

The title of the incident.

newIncidentRecord

Arguments

:: Text
$sel:arn:IncidentRecord'

-> UTCTime
$sel:creationTime:IncidentRecord'

-> Text
$sel:dedupeString:IncidentRecord'

-> Natural
$sel:impact:IncidentRecord'

-> IncidentRecordSource
$sel:incidentRecordSource:IncidentRecord'

-> Text
$sel:lastModifiedBy:IncidentRecord'

-> UTCTime
$sel:lastModifiedTime:IncidentRecord'

-> IncidentRecordStatus
$sel:status:IncidentRecord'

-> Text
$sel:title:IncidentRecord'

-> IncidentRecord

Create a value of IncidentRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:IncidentRecord', incidentRecord_summary - The summary of the incident. The summary is a brief synopsis of what occurred, what is currently happening, and context.

$sel:notificationTargets:IncidentRecord', incidentRecord_notificationTargets - The SNS targets that are notified when updates are made to an incident.

$sel:resolvedTime:IncidentRecord', incidentRecord_resolvedTime - The time at which the incident was resolved. This appears as a timeline event.

$sel:chatChannel:IncidentRecord', incidentRecord_chatChannel - The chat channel used for collaboration during an incident.

$sel:automationExecutions:IncidentRecord', incidentRecord_automationExecutions - The runbook, or automation document, that's run at the beginning of the incident.

$sel:arn:IncidentRecord', incidentRecord_arn - The Amazon Resource Name (ARN) of the incident record.

$sel:creationTime:IncidentRecord', incidentRecord_creationTime - The time that Incident Manager created the incident record.

$sel:dedupeString:IncidentRecord', incidentRecord_dedupeString - The string Incident Manager uses to prevent duplicate incidents from being created by the same incident.

$sel:impact:IncidentRecord', incidentRecord_impact - The impact of the incident on customers and applications.

$sel:incidentRecordSource:IncidentRecord', incidentRecord_incidentRecordSource - Details about the action that started the incident.

$sel:lastModifiedBy:IncidentRecord', incidentRecord_lastModifiedBy - Who modified the incident most recently.

$sel:lastModifiedTime:IncidentRecord', incidentRecord_lastModifiedTime - The time at which the incident was most recently modified.

$sel:status:IncidentRecord', incidentRecord_status - The current status of the incident.

$sel:title:IncidentRecord', incidentRecord_title - The title of the incident.

incidentRecord_summary :: Lens' IncidentRecord (Maybe Text)

The summary of the incident. The summary is a brief synopsis of what occurred, what is currently happening, and context.

incidentRecord_notificationTargets :: Lens' IncidentRecord (Maybe [NotificationTargetItem])

The SNS targets that are notified when updates are made to an incident.

incidentRecord_resolvedTime :: Lens' IncidentRecord (Maybe UTCTime)

The time at which the incident was resolved. This appears as a timeline event.

incidentRecord_chatChannel :: Lens' IncidentRecord (Maybe ChatChannel)

The chat channel used for collaboration during an incident.

incidentRecord_automationExecutions :: Lens' IncidentRecord (Maybe [AutomationExecution])

The runbook, or automation document, that's run at the beginning of the incident.

incidentRecord_arn :: Lens' IncidentRecord Text

The Amazon Resource Name (ARN) of the incident record.

incidentRecord_creationTime :: Lens' IncidentRecord UTCTime

The time that Incident Manager created the incident record.

incidentRecord_dedupeString :: Lens' IncidentRecord Text

The string Incident Manager uses to prevent duplicate incidents from being created by the same incident.

incidentRecord_impact :: Lens' IncidentRecord Natural

The impact of the incident on customers and applications.

incidentRecord_incidentRecordSource :: Lens' IncidentRecord IncidentRecordSource

Details about the action that started the incident.

incidentRecord_lastModifiedBy :: Lens' IncidentRecord Text

Who modified the incident most recently.

incidentRecord_lastModifiedTime :: Lens' IncidentRecord UTCTime

The time at which the incident was most recently modified.

incidentRecord_status :: Lens' IncidentRecord IncidentRecordStatus

The current status of the incident.

incidentRecord_title :: Lens' IncidentRecord Text

The title of the incident.