Amazonka.SSMIncidents.Types.IncidentRecordSource

Description

Documentation

data IncidentRecordSource

Details about how the incident record was created and when.

See: newIncidentRecordSource smart constructor.

Constructors

IncidentRecordSource'

Fields

invokedBy :: Maybe Text

The principal the assumed the role specified of the createdBy.

resourceArn :: Maybe Text

The resource that caused the incident to be created.

createdBy :: Text

The principal that started the incident.

source :: Text

The service that started the incident. This can be manually created from Incident Manager, automatically created using an AWS CloudWatch alarm, or Amazon EventBridge event.

newIncidentRecordSource

Arguments

:: Text
$sel:createdBy:IncidentRecordSource'

-> Text
$sel:source:IncidentRecordSource'

-> IncidentRecordSource

Create a value of IncidentRecordSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invokedBy:IncidentRecordSource', incidentRecordSource_invokedBy - The principal the assumed the role specified of the createdBy.

$sel:resourceArn:IncidentRecordSource', incidentRecordSource_resourceArn - The resource that caused the incident to be created.

$sel:createdBy:IncidentRecordSource', incidentRecordSource_createdBy - The principal that started the incident.

$sel:source:IncidentRecordSource', incidentRecordSource_source - The service that started the incident. This can be manually created from Incident Manager, automatically created using an AWS CloudWatch alarm, or Amazon EventBridge event.

incidentRecordSource_invokedBy :: Lens' IncidentRecordSource (Maybe Text)

The principal the assumed the role specified of the createdBy.

incidentRecordSource_resourceArn :: Lens' IncidentRecordSource (Maybe Text)

The resource that caused the incident to be created.

incidentRecordSource_createdBy :: Lens' IncidentRecordSource Text

The principal that started the incident.

incidentRecordSource_source :: Lens' IncidentRecordSource Text

The service that started the incident. This can be manually created from Incident Manager, automatically created using an AWS CloudWatch alarm, or Amazon EventBridge event.