Amazonka.SSMIncidents.Types.TimelineEvent

Description

Documentation

data TimelineEvent

A significant event that happened during the incident.

See: newTimelineEvent smart constructor.

Constructors

TimelineEvent'

Fields

eventData :: Text

A short description of the event.

eventId :: Text

The ID of the timeline event.

eventTime :: POSIX

The time that the event occurred.

eventType :: Text

The type of event that occurred. Currently Incident Manager supports only the Custom Event type.

eventUpdatedTime :: POSIX

The time that the timeline event was last updated.

incidentRecordArn :: Text

The Amazon Resource Name (ARN) of the incident that the event occurred during.

newTimelineEvent

Arguments

:: Text
$sel:eventData:TimelineEvent'

-> Text
$sel:eventId:TimelineEvent'

-> UTCTime
$sel:eventTime:TimelineEvent'

-> Text
$sel:eventType:TimelineEvent'

-> UTCTime
$sel:eventUpdatedTime:TimelineEvent'

-> Text
$sel:incidentRecordArn:TimelineEvent'

-> TimelineEvent

Create a value of TimelineEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventData:TimelineEvent', timelineEvent_eventData - A short description of the event.

$sel:eventId:TimelineEvent', timelineEvent_eventId - The ID of the timeline event.

$sel:eventTime:TimelineEvent', timelineEvent_eventTime - The time that the event occurred.

$sel:eventType:TimelineEvent', timelineEvent_eventType - The type of event that occurred. Currently Incident Manager supports only the Custom Event type.

$sel:eventUpdatedTime:TimelineEvent', timelineEvent_eventUpdatedTime - The time that the timeline event was last updated.

$sel:incidentRecordArn:TimelineEvent', timelineEvent_incidentRecordArn - The Amazon Resource Name (ARN) of the incident that the event occurred during.

timelineEvent_eventData :: Lens' TimelineEvent Text

A short description of the event.

timelineEvent_eventId :: Lens' TimelineEvent Text

The ID of the timeline event.

timelineEvent_eventTime :: Lens' TimelineEvent UTCTime

The time that the event occurred.

timelineEvent_eventType :: Lens' TimelineEvent Text

The type of event that occurred. Currently Incident Manager supports only the Custom Event type.

timelineEvent_eventUpdatedTime :: Lens' TimelineEvent UTCTime

The time that the timeline event was last updated.

timelineEvent_incidentRecordArn :: Lens' TimelineEvent Text

The Amazon Resource Name (ARN) of the incident that the event occurred during.