Amazonka.SSMIncidents.UpdateIncidentRecord

Description

Update the details of an incident record. You can use this action to update an incident record from the defined chat channel. For more information about using actions in chat channels, see Interacting through chat.

Creating a Request

data UpdateIncidentRecord

See: newUpdateIncidentRecord smart constructor.

Constructors

UpdateIncidentRecord'

Fields

summary :: Maybe Text

The summary describes what has happened during the incident.

status :: Maybe IncidentRecordStatus

The status of the incident. An incident can be Open or Resolved.

notificationTargets :: Maybe [NotificationTargetItem]

The SNS targets that are notified when updates are made to an incident.

Using multiple SNS topics creates redundancy in the case that a Region is down during the incident.

clientToken :: Maybe Text

A token ensuring that the action is called only once with the specified details.

impact :: Maybe Natural

Defines the impact to customers and applications. Providing an impact overwrites the impact provided by the response plan.

Possible impacts:

1 - Critical impact, this typically relates to full application failure that impacts many to all customers.

2 - High impact, partial application failure with impact to many customers.

3 - Medium impact, the application is providing reduced service to customers.

4 - Low impact, customer might aren't impacted by the problem yet.

5 - No impact, customers aren't currently impacted but urgent action is needed to avoid impact.

chatChannel :: Maybe ChatChannel

The AWS Chatbot chat channel for responders to collaborate in.

title :: Maybe Text

The title of the incident is a brief and easily recognizable.

arn :: Text

The Amazon Resource Name (ARN) of the incident record you are updating.

newUpdateIncidentRecord

Arguments

:: Text
$sel:arn:UpdateIncidentRecord'

-> UpdateIncidentRecord

Create a value of UpdateIncidentRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:UpdateIncidentRecord', updateIncidentRecord_summary - The summary describes what has happened during the incident.

$sel:status:UpdateIncidentRecord', updateIncidentRecord_status - The status of the incident. An incident can be Open or Resolved.

$sel:notificationTargets:UpdateIncidentRecord', updateIncidentRecord_notificationTargets - The SNS targets that are notified when updates are made to an incident.

Using multiple SNS topics creates redundancy in the case that a Region is down during the incident.

$sel:clientToken:UpdateIncidentRecord', updateIncidentRecord_clientToken - A token ensuring that the action is called only once with the specified details.

$sel:impact:UpdateIncidentRecord', updateIncidentRecord_impact - Defines the impact to customers and applications. Providing an impact overwrites the impact provided by the response plan.

Possible impacts:

1 - Critical impact, this typically relates to full application failure that impacts many to all customers.

2 - High impact, partial application failure with impact to many customers.

3 - Medium impact, the application is providing reduced service to customers.

4 - Low impact, customer might aren't impacted by the problem yet.

5 - No impact, customers aren't currently impacted but urgent action is needed to avoid impact.

$sel:chatChannel:UpdateIncidentRecord', updateIncidentRecord_chatChannel - The AWS Chatbot chat channel for responders to collaborate in.

$sel:title:UpdateIncidentRecord', updateIncidentRecord_title - The title of the incident is a brief and easily recognizable.

$sel:arn:UpdateIncidentRecord', updateIncidentRecord_arn - The Amazon Resource Name (ARN) of the incident record you are updating.

Request Lenses

updateIncidentRecord_summary :: Lens' UpdateIncidentRecord (Maybe Text)

The summary describes what has happened during the incident.

updateIncidentRecord_status :: Lens' UpdateIncidentRecord (Maybe IncidentRecordStatus)

The status of the incident. An incident can be Open or Resolved.

updateIncidentRecord_notificationTargets :: Lens' UpdateIncidentRecord (Maybe [NotificationTargetItem])

The SNS targets that are notified when updates are made to an incident.

Using multiple SNS topics creates redundancy in the case that a Region is down during the incident.

updateIncidentRecord_clientToken :: Lens' UpdateIncidentRecord (Maybe Text)

A token ensuring that the action is called only once with the specified details.

updateIncidentRecord_impact :: Lens' UpdateIncidentRecord (Maybe Natural)

Defines the impact to customers and applications. Providing an impact overwrites the impact provided by the response plan.

Possible impacts:

1 - Critical impact, this typically relates to full application failure that impacts many to all customers.

2 - High impact, partial application failure with impact to many customers.

3 - Medium impact, the application is providing reduced service to customers.

4 - Low impact, customer might aren't impacted by the problem yet.

5 - No impact, customers aren't currently impacted but urgent action is needed to avoid impact.

updateIncidentRecord_chatChannel :: Lens' UpdateIncidentRecord (Maybe ChatChannel)

The AWS Chatbot chat channel for responders to collaborate in.

updateIncidentRecord_title :: Lens' UpdateIncidentRecord (Maybe Text)

The title of the incident is a brief and easily recognizable.

updateIncidentRecord_arn :: Lens' UpdateIncidentRecord Text

The Amazon Resource Name (ARN) of the incident record you are updating.

Destructuring the Response

data UpdateIncidentRecordResponse

See: newUpdateIncidentRecordResponse smart constructor.

Constructors

UpdateIncidentRecordResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateIncidentRecordResponse

Arguments

:: Int
$sel:httpStatus:UpdateIncidentRecordResponse'

-> UpdateIncidentRecordResponse

Create a value of UpdateIncidentRecordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateIncidentRecordResponse', updateIncidentRecordResponse_httpStatus - The response's http status code.

Response Lenses

updateIncidentRecordResponse_httpStatus :: Lens' UpdateIncidentRecordResponse Int

The response's http status code.