Amazonka.SSO.Types

Description

Service Configuration

defaultService :: Service

API version 2019-06-10 of the Amazon Single Sign-On SDK configuration.

Errors

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a problem occurred with the input to the request. For example, a required parameter might be missing or out of range.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the request is being made too frequently and is more than what the server can handle.

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the request is not authorized. This can happen due to an invalid access token in the request.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource doesn't exist.

AccountInfo

data AccountInfo

Provides information about your AWS account.

See: newAccountInfo smart constructor.

Constructors

AccountInfo'

Fields

accountName :: Maybe Text

The display name of the AWS account that is assigned to the user.

accountId :: Maybe Text

The identifier of the AWS account that is assigned to the user.

emailAddress :: Maybe Text

The email address of the AWS account that is assigned to the user.

newAccountInfo :: AccountInfo

Create a value of AccountInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountName:AccountInfo', accountInfo_accountName - The display name of the AWS account that is assigned to the user.

$sel:accountId:AccountInfo', accountInfo_accountId - The identifier of the AWS account that is assigned to the user.

$sel:emailAddress:AccountInfo', accountInfo_emailAddress - The email address of the AWS account that is assigned to the user.

accountInfo_accountName :: Lens' AccountInfo (Maybe Text)

The display name of the AWS account that is assigned to the user.

accountInfo_accountId :: Lens' AccountInfo (Maybe Text)

The identifier of the AWS account that is assigned to the user.

accountInfo_emailAddress :: Lens' AccountInfo (Maybe Text)

The email address of the AWS account that is assigned to the user.

RoleCredentials

data RoleCredentials

Provides information about the role credentials that are assigned to the user.

See: newRoleCredentials smart constructor.

Constructors

RoleCredentials'

Fields

secretAccessKey :: Maybe (Sensitive Text)

The key that is used to sign the request. For more information, see Using Temporary Security Credentials to Request Access to AWS Resources in the AWS IAM User Guide.

sessionToken :: Maybe (Sensitive Text)

The token used for temporary credentials. For more information, see Using Temporary Security Credentials to Request Access to AWS Resources in the AWS IAM User Guide.

expiration :: Maybe Integer

The date on which temporary security credentials expire.

accessKeyId :: Maybe Text

The identifier used for the temporary security credentials. For more information, see Using Temporary Security Credentials to Request Access to AWS Resources in the AWS IAM User Guide.

newRoleCredentials :: RoleCredentials

Create a value of RoleCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secretAccessKey:RoleCredentials', roleCredentials_secretAccessKey - The key that is used to sign the request. For more information, see Using Temporary Security Credentials to Request Access to AWS Resources in the AWS IAM User Guide.

$sel:sessionToken:RoleCredentials', roleCredentials_sessionToken - The token used for temporary credentials. For more information, see Using Temporary Security Credentials to Request Access to AWS Resources in the AWS IAM User Guide.

$sel:expiration:RoleCredentials', roleCredentials_expiration - The date on which temporary security credentials expire.

$sel:accessKeyId:RoleCredentials', roleCredentials_accessKeyId - The identifier used for the temporary security credentials. For more information, see Using Temporary Security Credentials to Request Access to AWS Resources in the AWS IAM User Guide.

roleCredentials_secretAccessKey :: Lens' RoleCredentials (Maybe Text)

The key that is used to sign the request. For more information, see Using Temporary Security Credentials to Request Access to AWS Resources in the AWS IAM User Guide.

roleCredentials_sessionToken :: Lens' RoleCredentials (Maybe Text)

The token used for temporary credentials. For more information, see Using Temporary Security Credentials to Request Access to AWS Resources in the AWS IAM User Guide.

roleCredentials_expiration :: Lens' RoleCredentials (Maybe Integer)

The date on which temporary security credentials expire.

roleCredentials_accessKeyId :: Lens' RoleCredentials (Maybe Text)

The identifier used for the temporary security credentials. For more information, see Using Temporary Security Credentials to Request Access to AWS Resources in the AWS IAM User Guide.

RoleInfo

data RoleInfo

Provides information about the role that is assigned to the user.

See: newRoleInfo smart constructor.

Constructors

RoleInfo'

Fields

roleName :: Maybe Text

The friendly name of the role that is assigned to the user.

accountId :: Maybe Text

The identifier of the AWS account assigned to the user.

newRoleInfo :: RoleInfo

Create a value of RoleInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleName:RoleInfo', roleInfo_roleName - The friendly name of the role that is assigned to the user.

$sel:accountId:RoleInfo', roleInfo_accountId - The identifier of the AWS account assigned to the user.

roleInfo_roleName :: Lens' RoleInfo (Maybe Text)

The friendly name of the role that is assigned to the user.

roleInfo_accountId :: Lens' RoleInfo (Maybe Text)

The identifier of the AWS account assigned to the user.