Amazonka.SSOAdmin.AttachManagedPolicyToPermissionSet

Description

Attaches an IAM managed policy ARN to a permission set.

If the permission set is already referenced by one or more account assignments, you will need to call ProvisionPermissionSet after this operation. Calling ProvisionPermissionSet applies the corresponding IAM policy updates to all assigned accounts.

Creating a Request

data AttachManagedPolicyToPermissionSet

See: newAttachManagedPolicyToPermissionSet smart constructor.

Constructors

AttachManagedPolicyToPermissionSet'

Fields

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

permissionSetArn :: Text

The ARN of the PermissionSet that the managed policy should be attached to.

managedPolicyArn :: Text

The IAM managed policy ARN to be attached to a permission set.

newAttachManagedPolicyToPermissionSet

Arguments

:: Text
$sel:instanceArn:AttachManagedPolicyToPermissionSet'

-> Text
$sel:permissionSetArn:AttachManagedPolicyToPermissionSet'

-> Text
$sel:managedPolicyArn:AttachManagedPolicyToPermissionSet'

-> AttachManagedPolicyToPermissionSet

Create a value of AttachManagedPolicyToPermissionSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceArn:AttachManagedPolicyToPermissionSet', attachManagedPolicyToPermissionSet_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

$sel:permissionSetArn:AttachManagedPolicyToPermissionSet', attachManagedPolicyToPermissionSet_permissionSetArn - The ARN of the PermissionSet that the managed policy should be attached to.

$sel:managedPolicyArn:AttachManagedPolicyToPermissionSet', attachManagedPolicyToPermissionSet_managedPolicyArn - The IAM managed policy ARN to be attached to a permission set.

Request Lenses

attachManagedPolicyToPermissionSet_instanceArn :: Lens' AttachManagedPolicyToPermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

attachManagedPolicyToPermissionSet_permissionSetArn :: Lens' AttachManagedPolicyToPermissionSet Text

The ARN of the PermissionSet that the managed policy should be attached to.

attachManagedPolicyToPermissionSet_managedPolicyArn :: Lens' AttachManagedPolicyToPermissionSet Text

The IAM managed policy ARN to be attached to a permission set.

Destructuring the Response

data AttachManagedPolicyToPermissionSetResponse

See: newAttachManagedPolicyToPermissionSetResponse smart constructor.

Constructors

AttachManagedPolicyToPermissionSetResponse'

Fields

httpStatus :: Int

The response's http status code.

newAttachManagedPolicyToPermissionSetResponse

Arguments

:: Int
$sel:httpStatus:AttachManagedPolicyToPermissionSetResponse'

-> AttachManagedPolicyToPermissionSetResponse

Create a value of AttachManagedPolicyToPermissionSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AttachManagedPolicyToPermissionSetResponse', attachManagedPolicyToPermissionSetResponse_httpStatus - The response's http status code.

Response Lenses

attachManagedPolicyToPermissionSetResponse_httpStatus :: Lens' AttachManagedPolicyToPermissionSetResponse Int

The response's http status code.