Amazonka.SSOAdmin.DeleteAccountAssignment

Description

Deletes a principal's access from a specified Amazon Web Services account using a specified permission set.

Creating a Request

data DeleteAccountAssignment

See: newDeleteAccountAssignment smart constructor.

Constructors

DeleteAccountAssignment'

Fields

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

targetId :: Text

TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

targetType :: TargetType

The entity type for which the assignment will be deleted.

permissionSetArn :: Text

The ARN of the permission set that will be used to remove access.

principalType :: PrincipalType

The entity type for which the assignment will be deleted.

principalId :: Text

An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

newDeleteAccountAssignment

Arguments

:: Text
$sel:instanceArn:DeleteAccountAssignment'

-> Text
$sel:targetId:DeleteAccountAssignment'

-> TargetType
$sel:targetType:DeleteAccountAssignment'

-> Text
$sel:permissionSetArn:DeleteAccountAssignment'

-> PrincipalType
$sel:principalType:DeleteAccountAssignment'

-> Text
$sel:principalId:DeleteAccountAssignment'

-> DeleteAccountAssignment

Create a value of DeleteAccountAssignment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceArn:DeleteAccountAssignment', deleteAccountAssignment_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

$sel:targetId:DeleteAccountAssignment', deleteAccountAssignment_targetId - TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

$sel:targetType:DeleteAccountAssignment', deleteAccountAssignment_targetType - The entity type for which the assignment will be deleted.

$sel:permissionSetArn:DeleteAccountAssignment', deleteAccountAssignment_permissionSetArn - The ARN of the permission set that will be used to remove access.

$sel:principalType:DeleteAccountAssignment', deleteAccountAssignment_principalType - The entity type for which the assignment will be deleted.

$sel:principalId:DeleteAccountAssignment', deleteAccountAssignment_principalId - An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

Request Lenses

deleteAccountAssignment_instanceArn :: Lens' DeleteAccountAssignment Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

deleteAccountAssignment_targetId :: Lens' DeleteAccountAssignment Text

TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

deleteAccountAssignment_targetType :: Lens' DeleteAccountAssignment TargetType

The entity type for which the assignment will be deleted.

deleteAccountAssignment_permissionSetArn :: Lens' DeleteAccountAssignment Text

The ARN of the permission set that will be used to remove access.

deleteAccountAssignment_principalType :: Lens' DeleteAccountAssignment PrincipalType

The entity type for which the assignment will be deleted.

deleteAccountAssignment_principalId :: Lens' DeleteAccountAssignment Text

An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

Destructuring the Response

data DeleteAccountAssignmentResponse

See: newDeleteAccountAssignmentResponse smart constructor.

Constructors

DeleteAccountAssignmentResponse'

Fields

accountAssignmentDeletionStatus :: Maybe AccountAssignmentOperationStatus

The status object for the account assignment deletion operation.

httpStatus :: Int

The response's http status code.

newDeleteAccountAssignmentResponse

Arguments

:: Int
$sel:httpStatus:DeleteAccountAssignmentResponse'

-> DeleteAccountAssignmentResponse

Create a value of DeleteAccountAssignmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountAssignmentDeletionStatus:DeleteAccountAssignmentResponse', deleteAccountAssignmentResponse_accountAssignmentDeletionStatus - The status object for the account assignment deletion operation.

$sel:httpStatus:DeleteAccountAssignmentResponse', deleteAccountAssignmentResponse_httpStatus - The response's http status code.

Response Lenses

deleteAccountAssignmentResponse_accountAssignmentDeletionStatus :: Lens' DeleteAccountAssignmentResponse (Maybe AccountAssignmentOperationStatus)

The status object for the account assignment deletion operation.

deleteAccountAssignmentResponse_httpStatus :: Lens' DeleteAccountAssignmentResponse Int

The response's http status code.