Amazonka.SSOAdmin.DescribeAccountAssignmentCreationStatus

Description

Describes the status of the assignment creation request.

Creating a Request

data DescribeAccountAssignmentCreationStatus

See: newDescribeAccountAssignmentCreationStatus smart constructor.

Constructors

DescribeAccountAssignmentCreationStatus'

Fields

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

accountAssignmentCreationRequestId :: Text

The identifier that is used to track the request operation progress.

newDescribeAccountAssignmentCreationStatus

Arguments

:: Text
$sel:instanceArn:DescribeAccountAssignmentCreationStatus'

-> Text
$sel:accountAssignmentCreationRequestId:DescribeAccountAssignmentCreationStatus'

-> DescribeAccountAssignmentCreationStatus

Create a value of DescribeAccountAssignmentCreationStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceArn:DescribeAccountAssignmentCreationStatus', describeAccountAssignmentCreationStatus_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

$sel:accountAssignmentCreationRequestId:DescribeAccountAssignmentCreationStatus', describeAccountAssignmentCreationStatus_accountAssignmentCreationRequestId - The identifier that is used to track the request operation progress.

Request Lenses

describeAccountAssignmentCreationStatus_instanceArn :: Lens' DescribeAccountAssignmentCreationStatus Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

describeAccountAssignmentCreationStatus_accountAssignmentCreationRequestId :: Lens' DescribeAccountAssignmentCreationStatus Text

The identifier that is used to track the request operation progress.

Destructuring the Response

data DescribeAccountAssignmentCreationStatusResponse

See: newDescribeAccountAssignmentCreationStatusResponse smart constructor.

Constructors

DescribeAccountAssignmentCreationStatusResponse'

Fields

accountAssignmentCreationStatus :: Maybe AccountAssignmentOperationStatus

The status object for the account assignment creation operation.

httpStatus :: Int

The response's http status code.

newDescribeAccountAssignmentCreationStatusResponse

Arguments

:: Int
$sel:httpStatus:DescribeAccountAssignmentCreationStatusResponse'

-> DescribeAccountAssignmentCreationStatusResponse

Create a value of DescribeAccountAssignmentCreationStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountAssignmentCreationStatus:DescribeAccountAssignmentCreationStatusResponse', describeAccountAssignmentCreationStatusResponse_accountAssignmentCreationStatus - The status object for the account assignment creation operation.

$sel:httpStatus:DescribeAccountAssignmentCreationStatusResponse', describeAccountAssignmentCreationStatusResponse_httpStatus - The response's http status code.

Response Lenses

describeAccountAssignmentCreationStatusResponse_accountAssignmentCreationStatus :: Lens' DescribeAccountAssignmentCreationStatusResponse (Maybe AccountAssignmentOperationStatus)

The status object for the account assignment creation operation.

describeAccountAssignmentCreationStatusResponse_httpStatus :: Lens' DescribeAccountAssignmentCreationStatusResponse Int

The response's http status code.