Amazonka.SSOAdmin.GetInlinePolicyForPermissionSet

Description

Obtains the inline policy assigned to the permission set.

Creating a Request

data GetInlinePolicyForPermissionSet

See: newGetInlinePolicyForPermissionSet smart constructor.

Constructors

GetInlinePolicyForPermissionSet'

Fields

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

permissionSetArn :: Text

The ARN of the permission set.

newGetInlinePolicyForPermissionSet

Arguments

:: Text
$sel:instanceArn:GetInlinePolicyForPermissionSet'

-> Text
$sel:permissionSetArn:GetInlinePolicyForPermissionSet'

-> GetInlinePolicyForPermissionSet

Create a value of GetInlinePolicyForPermissionSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceArn:GetInlinePolicyForPermissionSet', getInlinePolicyForPermissionSet_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

$sel:permissionSetArn:GetInlinePolicyForPermissionSet', getInlinePolicyForPermissionSet_permissionSetArn - The ARN of the permission set.

Request Lenses

getInlinePolicyForPermissionSet_instanceArn :: Lens' GetInlinePolicyForPermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

getInlinePolicyForPermissionSet_permissionSetArn :: Lens' GetInlinePolicyForPermissionSet Text

The ARN of the permission set.

Destructuring the Response

data GetInlinePolicyForPermissionSetResponse

See: newGetInlinePolicyForPermissionSetResponse smart constructor.

Constructors

GetInlinePolicyForPermissionSetResponse'

Fields

inlinePolicy :: Maybe (Sensitive Text)

The IAM inline policy that is attached to the permission set.

httpStatus :: Int

The response's http status code.

newGetInlinePolicyForPermissionSetResponse

Arguments

:: Int
$sel:httpStatus:GetInlinePolicyForPermissionSetResponse'

-> GetInlinePolicyForPermissionSetResponse

Create a value of GetInlinePolicyForPermissionSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inlinePolicy:GetInlinePolicyForPermissionSetResponse', getInlinePolicyForPermissionSetResponse_inlinePolicy - The IAM inline policy that is attached to the permission set.

$sel:httpStatus:GetInlinePolicyForPermissionSetResponse', getInlinePolicyForPermissionSetResponse_httpStatus - The response's http status code.

Response Lenses

getInlinePolicyForPermissionSetResponse_inlinePolicy :: Lens' GetInlinePolicyForPermissionSetResponse (Maybe Text)

The IAM inline policy that is attached to the permission set.

getInlinePolicyForPermissionSetResponse_httpStatus :: Lens' GetInlinePolicyForPermissionSetResponse Int

The response's http status code.