Amazonka.SSOAdmin.Lens

Description

Operations

DescribePermissionSet

describePermissionSet_instanceArn :: Lens' DescribePermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

describePermissionSet_permissionSetArn :: Lens' DescribePermissionSet Text

The ARN of the permission set.

describePermissionSetResponse_permissionSet :: Lens' DescribePermissionSetResponse (Maybe PermissionSet)

Describes the level of access on an Amazon Web Services account.

describePermissionSetResponse_httpStatus :: Lens' DescribePermissionSetResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_nextToken :: Lens' ListTagsForResource (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listTagsForResource_instanceArn :: Lens' ListTagsForResource Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the resource with the tags to be listed.

listTagsForResourceResponse_nextToken :: Lens' ListTagsForResourceResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

A set of key-value pairs that are used to manage the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CreateInstanceAccessControlAttributeConfiguration

createInstanceAccessControlAttributeConfiguration_instanceArn :: Lens' CreateInstanceAccessControlAttributeConfiguration Text

The ARN of the SSO instance under which the operation will be executed.

createInstanceAccessControlAttributeConfiguration_instanceAccessControlAttributeConfiguration :: Lens' CreateInstanceAccessControlAttributeConfiguration InstanceAccessControlAttributeConfiguration

Specifies the Amazon Web Services SSO identity store attributes to add to your ABAC configuration. When using an external identity provider as an identity source, you can pass attributes through the SAML assertion. Doing so provides an alternative to configuring attributes from the Amazon Web Services SSO identity store. If a SAML assertion passes any of these attributes, Amazon Web Services SSO will replace the attribute value with the value from the Amazon Web Services SSO identity store.

createInstanceAccessControlAttributeConfigurationResponse_httpStatus :: Lens' CreateInstanceAccessControlAttributeConfigurationResponse Int

The response's http status code.

GetInlinePolicyForPermissionSet

getInlinePolicyForPermissionSet_instanceArn :: Lens' GetInlinePolicyForPermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

getInlinePolicyForPermissionSet_permissionSetArn :: Lens' GetInlinePolicyForPermissionSet Text

The ARN of the permission set.

getInlinePolicyForPermissionSetResponse_inlinePolicy :: Lens' GetInlinePolicyForPermissionSetResponse (Maybe Text)

The IAM inline policy that is attached to the permission set.

getInlinePolicyForPermissionSetResponse_httpStatus :: Lens' GetInlinePolicyForPermissionSetResponse Int

The response's http status code.

ListPermissionSets

listPermissionSets_nextToken :: Lens' ListPermissionSets (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listPermissionSets_maxResults :: Lens' ListPermissionSets (Maybe Natural)

The maximum number of results to display for the assignment.

listPermissionSets_instanceArn :: Lens' ListPermissionSets Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

listPermissionSetsResponse_permissionSets :: Lens' ListPermissionSetsResponse (Maybe [Text])

Defines the level of access on an Amazon Web Services account.

listPermissionSetsResponse_nextToken :: Lens' ListPermissionSetsResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listPermissionSetsResponse_httpStatus :: Lens' ListPermissionSetsResponse Int

The response's http status code.

DeletePermissionSet

deletePermissionSet_instanceArn :: Lens' DeletePermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

deletePermissionSet_permissionSetArn :: Lens' DeletePermissionSet Text

The ARN of the permission set that should be deleted.

deletePermissionSetResponse_httpStatus :: Lens' DeletePermissionSetResponse Int

The response's http status code.

UpdatePermissionSet

updatePermissionSet_relayState :: Lens' UpdatePermissionSet (Maybe Text)

Used to redirect users within the application during the federation authentication process.

updatePermissionSet_sessionDuration :: Lens' UpdatePermissionSet (Maybe Text)

The length of time that the application user sessions are valid for in the ISO-8601 standard.

updatePermissionSet_description :: Lens' UpdatePermissionSet (Maybe Text)

The description of the PermissionSet.

updatePermissionSet_instanceArn :: Lens' UpdatePermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

updatePermissionSet_permissionSetArn :: Lens' UpdatePermissionSet Text

The ARN of the permission set.

updatePermissionSetResponse_httpStatus :: Lens' UpdatePermissionSetResponse Int

The response's http status code.

ProvisionPermissionSet

provisionPermissionSet_targetId :: Lens' ProvisionPermissionSet (Maybe Text)

TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

provisionPermissionSet_instanceArn :: Lens' ProvisionPermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

provisionPermissionSet_permissionSetArn :: Lens' ProvisionPermissionSet Text

The ARN of the permission set.

provisionPermissionSet_targetType :: Lens' ProvisionPermissionSet ProvisionTargetType

The entity type for which the assignment will be created.

provisionPermissionSetResponse_permissionSetProvisioningStatus :: Lens' ProvisionPermissionSetResponse (Maybe PermissionSetProvisioningStatus)

The status object for the permission set provisioning operation.

provisionPermissionSetResponse_httpStatus :: Lens' ProvisionPermissionSetResponse Int

The response's http status code.

ListAccountAssignments

listAccountAssignments_nextToken :: Lens' ListAccountAssignments (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listAccountAssignments_maxResults :: Lens' ListAccountAssignments (Maybe Natural)

The maximum number of results to display for the assignment.

listAccountAssignments_instanceArn :: Lens' ListAccountAssignments Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

listAccountAssignments_accountId :: Lens' ListAccountAssignments Text

The identifier of the Amazon Web Services account from which to list the assignments.

listAccountAssignments_permissionSetArn :: Lens' ListAccountAssignments Text

The ARN of the permission set from which to list assignments.

listAccountAssignmentsResponse_accountAssignments :: Lens' ListAccountAssignmentsResponse (Maybe [AccountAssignment])

The list of assignments that match the input Amazon Web Services account and permission set.

listAccountAssignmentsResponse_nextToken :: Lens' ListAccountAssignmentsResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listAccountAssignmentsResponse_httpStatus :: Lens' ListAccountAssignmentsResponse Int

The response's http status code.

DescribePermissionSetProvisioningStatus

describePermissionSetProvisioningStatus_instanceArn :: Lens' DescribePermissionSetProvisioningStatus Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

describePermissionSetProvisioningStatus_provisionPermissionSetRequestId :: Lens' DescribePermissionSetProvisioningStatus Text

The identifier that is provided by the ProvisionPermissionSet call to retrieve the current status of the provisioning workflow.

describePermissionSetProvisioningStatusResponse_permissionSetProvisioningStatus :: Lens' DescribePermissionSetProvisioningStatusResponse (Maybe PermissionSetProvisioningStatus)

The status object for the permission set provisioning operation.

describePermissionSetProvisioningStatusResponse_httpStatus :: Lens' DescribePermissionSetProvisioningStatusResponse Int

The response's http status code.

AttachManagedPolicyToPermissionSet

attachManagedPolicyToPermissionSet_instanceArn :: Lens' AttachManagedPolicyToPermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

attachManagedPolicyToPermissionSet_permissionSetArn :: Lens' AttachManagedPolicyToPermissionSet Text

The ARN of the PermissionSet that the managed policy should be attached to.

attachManagedPolicyToPermissionSet_managedPolicyArn :: Lens' AttachManagedPolicyToPermissionSet Text

The IAM managed policy ARN to be attached to a permission set.

attachManagedPolicyToPermissionSetResponse_httpStatus :: Lens' AttachManagedPolicyToPermissionSetResponse Int

The response's http status code.

ListManagedPoliciesInPermissionSet

listManagedPoliciesInPermissionSet_nextToken :: Lens' ListManagedPoliciesInPermissionSet (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listManagedPoliciesInPermissionSet_maxResults :: Lens' ListManagedPoliciesInPermissionSet (Maybe Natural)

The maximum number of results to display for the PermissionSet.

listManagedPoliciesInPermissionSet_instanceArn :: Lens' ListManagedPoliciesInPermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

listManagedPoliciesInPermissionSet_permissionSetArn :: Lens' ListManagedPoliciesInPermissionSet Text

The ARN of the PermissionSet whose managed policies will be listed.

listManagedPoliciesInPermissionSetResponse_nextToken :: Lens' ListManagedPoliciesInPermissionSetResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listManagedPoliciesInPermissionSetResponse_attachedManagedPolicies :: Lens' ListManagedPoliciesInPermissionSetResponse (Maybe [AttachedManagedPolicy])

The array of the AttachedManagedPolicy data type object.

listManagedPoliciesInPermissionSetResponse_httpStatus :: Lens' ListManagedPoliciesInPermissionSetResponse Int

The response's http status code.

DeleteInlinePolicyFromPermissionSet

deleteInlinePolicyFromPermissionSet_instanceArn :: Lens' DeleteInlinePolicyFromPermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

deleteInlinePolicyFromPermissionSet_permissionSetArn :: Lens' DeleteInlinePolicyFromPermissionSet Text

The ARN of the permission set that will be used to remove access.

deleteInlinePolicyFromPermissionSetResponse_httpStatus :: Lens' DeleteInlinePolicyFromPermissionSetResponse Int

The response's http status code.

DeleteInstanceAccessControlAttributeConfiguration

deleteInstanceAccessControlAttributeConfiguration_instanceArn :: Lens' DeleteInstanceAccessControlAttributeConfiguration Text

The ARN of the SSO instance under which the operation will be executed.

deleteInstanceAccessControlAttributeConfigurationResponse_httpStatus :: Lens' DeleteInstanceAccessControlAttributeConfigurationResponse Int

The response's http status code.

UpdateInstanceAccessControlAttributeConfiguration

updateInstanceAccessControlAttributeConfiguration_instanceArn :: Lens' UpdateInstanceAccessControlAttributeConfiguration Text

The ARN of the SSO instance under which the operation will be executed.

updateInstanceAccessControlAttributeConfiguration_instanceAccessControlAttributeConfiguration :: Lens' UpdateInstanceAccessControlAttributeConfiguration InstanceAccessControlAttributeConfiguration

Updates the attributes for your ABAC configuration.

updateInstanceAccessControlAttributeConfigurationResponse_httpStatus :: Lens' UpdateInstanceAccessControlAttributeConfigurationResponse Int

The response's http status code.

DescribeAccountAssignmentDeletionStatus

describeAccountAssignmentDeletionStatus_instanceArn :: Lens' DescribeAccountAssignmentDeletionStatus Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

describeAccountAssignmentDeletionStatus_accountAssignmentDeletionRequestId :: Lens' DescribeAccountAssignmentDeletionStatus Text

The identifier that is used to track the request operation progress.

describeAccountAssignmentDeletionStatusResponse_accountAssignmentDeletionStatus :: Lens' DescribeAccountAssignmentDeletionStatusResponse (Maybe AccountAssignmentOperationStatus)

The status object for the account assignment deletion operation.

describeAccountAssignmentDeletionStatusResponse_httpStatus :: Lens' DescribeAccountAssignmentDeletionStatusResponse Int

The response's http status code.

DescribeAccountAssignmentCreationStatus

describeAccountAssignmentCreationStatus_instanceArn :: Lens' DescribeAccountAssignmentCreationStatus Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

describeAccountAssignmentCreationStatus_accountAssignmentCreationRequestId :: Lens' DescribeAccountAssignmentCreationStatus Text

The identifier that is used to track the request operation progress.

describeAccountAssignmentCreationStatusResponse_accountAssignmentCreationStatus :: Lens' DescribeAccountAssignmentCreationStatusResponse (Maybe AccountAssignmentOperationStatus)

The status object for the account assignment creation operation.

describeAccountAssignmentCreationStatusResponse_httpStatus :: Lens' DescribeAccountAssignmentCreationStatusResponse Int

The response's http status code.

PutInlinePolicyToPermissionSet

putInlinePolicyToPermissionSet_instanceArn :: Lens' PutInlinePolicyToPermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

putInlinePolicyToPermissionSet_permissionSetArn :: Lens' PutInlinePolicyToPermissionSet Text

The ARN of the permission set.

putInlinePolicyToPermissionSet_inlinePolicy :: Lens' PutInlinePolicyToPermissionSet Text

The IAM inline policy to attach to a PermissionSet.

putInlinePolicyToPermissionSetResponse_httpStatus :: Lens' PutInlinePolicyToPermissionSetResponse Int

The response's http status code.

ListAccountsForProvisionedPermissionSet

listAccountsForProvisionedPermissionSet_provisioningStatus :: Lens' ListAccountsForProvisionedPermissionSet (Maybe ProvisioningStatus)

The permission set provisioning status for an Amazon Web Services account.

listAccountsForProvisionedPermissionSet_nextToken :: Lens' ListAccountsForProvisionedPermissionSet (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listAccountsForProvisionedPermissionSet_maxResults :: Lens' ListAccountsForProvisionedPermissionSet (Maybe Natural)

The maximum number of results to display for the PermissionSet.

listAccountsForProvisionedPermissionSet_instanceArn :: Lens' ListAccountsForProvisionedPermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

listAccountsForProvisionedPermissionSet_permissionSetArn :: Lens' ListAccountsForProvisionedPermissionSet Text

The ARN of the PermissionSet from which the associated Amazon Web Services accounts will be listed.

listAccountsForProvisionedPermissionSetResponse_accountIds :: Lens' ListAccountsForProvisionedPermissionSetResponse (Maybe [Text])

The list of Amazon Web Services AccountIds.

listAccountsForProvisionedPermissionSetResponse_nextToken :: Lens' ListAccountsForProvisionedPermissionSetResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listAccountsForProvisionedPermissionSetResponse_httpStatus :: Lens' ListAccountsForProvisionedPermissionSetResponse Int

The response's http status code.

ListPermissionSetsProvisionedToAccount

listPermissionSetsProvisionedToAccount_provisioningStatus :: Lens' ListPermissionSetsProvisionedToAccount (Maybe ProvisioningStatus)

The status object for the permission set provisioning operation.

listPermissionSetsProvisionedToAccount_nextToken :: Lens' ListPermissionSetsProvisionedToAccount (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listPermissionSetsProvisionedToAccount_maxResults :: Lens' ListPermissionSetsProvisionedToAccount (Maybe Natural)

The maximum number of results to display for the assignment.

listPermissionSetsProvisionedToAccount_instanceArn :: Lens' ListPermissionSetsProvisionedToAccount Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

listPermissionSetsProvisionedToAccount_accountId :: Lens' ListPermissionSetsProvisionedToAccount Text

The identifier of the Amazon Web Services account from which to list the assignments.

listPermissionSetsProvisionedToAccountResponse_permissionSets :: Lens' ListPermissionSetsProvisionedToAccountResponse (Maybe [Text])

Defines the level of access that an Amazon Web Services account has.

listPermissionSetsProvisionedToAccountResponse_nextToken :: Lens' ListPermissionSetsProvisionedToAccountResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listPermissionSetsProvisionedToAccountResponse_httpStatus :: Lens' ListPermissionSetsProvisionedToAccountResponse Int

The response's http status code.

DetachManagedPolicyFromPermissionSet

detachManagedPolicyFromPermissionSet_instanceArn :: Lens' DetachManagedPolicyFromPermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

detachManagedPolicyFromPermissionSet_permissionSetArn :: Lens' DetachManagedPolicyFromPermissionSet Text

The ARN of the PermissionSet from which the policy should be detached.

detachManagedPolicyFromPermissionSet_managedPolicyArn :: Lens' DetachManagedPolicyFromPermissionSet Text

The IAM managed policy ARN to be attached to a permission set.

detachManagedPolicyFromPermissionSetResponse_httpStatus :: Lens' DetachManagedPolicyFromPermissionSetResponse Int

The response's http status code.

ListAccountAssignmentCreationStatus

listAccountAssignmentCreationStatus_nextToken :: Lens' ListAccountAssignmentCreationStatus (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listAccountAssignmentCreationStatus_filter :: Lens' ListAccountAssignmentCreationStatus (Maybe OperationStatusFilter)

Filters results based on the passed attribute value.

listAccountAssignmentCreationStatus_maxResults :: Lens' ListAccountAssignmentCreationStatus (Maybe Natural)

The maximum number of results to display for the assignment.

listAccountAssignmentCreationStatus_instanceArn :: Lens' ListAccountAssignmentCreationStatus Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

listAccountAssignmentCreationStatusResponse_accountAssignmentsCreationStatus :: Lens' ListAccountAssignmentCreationStatusResponse (Maybe [AccountAssignmentOperationStatusMetadata])

The status object for the account assignment creation operation.

listAccountAssignmentCreationStatusResponse_nextToken :: Lens' ListAccountAssignmentCreationStatusResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listAccountAssignmentCreationStatusResponse_httpStatus :: Lens' ListAccountAssignmentCreationStatusResponse Int

The response's http status code.

CreatePermissionSet

createPermissionSet_relayState :: Lens' CreatePermissionSet (Maybe Text)

Used to redirect users within the application during the federation authentication process.

createPermissionSet_sessionDuration :: Lens' CreatePermissionSet (Maybe Text)

The length of time that the application user sessions are valid in the ISO-8601 standard.

createPermissionSet_description :: Lens' CreatePermissionSet (Maybe Text)

The description of the PermissionSet.

createPermissionSet_tags :: Lens' CreatePermissionSet (Maybe [Tag])

The tags to attach to the new PermissionSet.

createPermissionSet_name :: Lens' CreatePermissionSet Text

The name of the PermissionSet.

createPermissionSet_instanceArn :: Lens' CreatePermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

createPermissionSetResponse_permissionSet :: Lens' CreatePermissionSetResponse (Maybe PermissionSet)

Defines the level of access on an Amazon Web Services account.

createPermissionSetResponse_httpStatus :: Lens' CreatePermissionSetResponse Int

The response's http status code.

ListAccountAssignmentDeletionStatus

listAccountAssignmentDeletionStatus_nextToken :: Lens' ListAccountAssignmentDeletionStatus (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listAccountAssignmentDeletionStatus_filter :: Lens' ListAccountAssignmentDeletionStatus (Maybe OperationStatusFilter)

Filters results based on the passed attribute value.

listAccountAssignmentDeletionStatus_maxResults :: Lens' ListAccountAssignmentDeletionStatus (Maybe Natural)

The maximum number of results to display for the assignment.

listAccountAssignmentDeletionStatus_instanceArn :: Lens' ListAccountAssignmentDeletionStatus Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

listAccountAssignmentDeletionStatusResponse_accountAssignmentsDeletionStatus :: Lens' ListAccountAssignmentDeletionStatusResponse (Maybe [AccountAssignmentOperationStatusMetadata])

The status object for the account assignment deletion operation.

listAccountAssignmentDeletionStatusResponse_nextToken :: Lens' ListAccountAssignmentDeletionStatusResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listAccountAssignmentDeletionStatusResponse_httpStatus :: Lens' ListAccountAssignmentDeletionStatusResponse Int

The response's http status code.

TagResource

tagResource_instanceArn :: Lens' TagResource Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the resource with the tags to be listed.

tagResource_tags :: Lens' TagResource [Tag]

A set of key-value pairs that are used to manage the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListInstances

listInstances_nextToken :: Lens' ListInstances (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listInstances_maxResults :: Lens' ListInstances (Maybe Natural)

The maximum number of results to display for the instance.

listInstancesResponse_nextToken :: Lens' ListInstancesResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listInstancesResponse_instances :: Lens' ListInstancesResponse (Maybe [InstanceMetadata])

Lists the SSO instances that the caller has access to.

listInstancesResponse_httpStatus :: Lens' ListInstancesResponse Int

The response's http status code.

DescribeInstanceAccessControlAttributeConfiguration

describeInstanceAccessControlAttributeConfiguration_instanceArn :: Lens' DescribeInstanceAccessControlAttributeConfiguration Text

The ARN of the SSO instance under which the operation will be executed.

describeInstanceAccessControlAttributeConfigurationResponse_status :: Lens' DescribeInstanceAccessControlAttributeConfigurationResponse (Maybe InstanceAccessControlAttributeConfigurationStatus)

The status of the attribute configuration process.

describeInstanceAccessControlAttributeConfigurationResponse_instanceAccessControlAttributeConfiguration :: Lens' DescribeInstanceAccessControlAttributeConfigurationResponse (Maybe InstanceAccessControlAttributeConfiguration)

Gets the list of Amazon Web Services SSO identity store attributes that have been added to your ABAC configuration.

describeInstanceAccessControlAttributeConfigurationResponse_statusReason :: Lens' DescribeInstanceAccessControlAttributeConfigurationResponse (Maybe Text)

Provides more details about the current status of the specified attribute.

describeInstanceAccessControlAttributeConfigurationResponse_httpStatus :: Lens' DescribeInstanceAccessControlAttributeConfigurationResponse Int

The response's http status code.

UntagResource

untagResource_instanceArn :: Lens' UntagResource Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the resource with the tags to be listed.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The keys of tags that are attached to the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DeleteAccountAssignment

deleteAccountAssignment_instanceArn :: Lens' DeleteAccountAssignment Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

deleteAccountAssignment_targetId :: Lens' DeleteAccountAssignment Text

TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

deleteAccountAssignment_targetType :: Lens' DeleteAccountAssignment TargetType

The entity type for which the assignment will be deleted.

deleteAccountAssignment_permissionSetArn :: Lens' DeleteAccountAssignment Text

The ARN of the permission set that will be used to remove access.

deleteAccountAssignment_principalType :: Lens' DeleteAccountAssignment PrincipalType

The entity type for which the assignment will be deleted.

deleteAccountAssignment_principalId :: Lens' DeleteAccountAssignment Text

An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

deleteAccountAssignmentResponse_accountAssignmentDeletionStatus :: Lens' DeleteAccountAssignmentResponse (Maybe AccountAssignmentOperationStatus)

The status object for the account assignment deletion operation.

deleteAccountAssignmentResponse_httpStatus :: Lens' DeleteAccountAssignmentResponse Int

The response's http status code.

ListPermissionSetProvisioningStatus

listPermissionSetProvisioningStatus_nextToken :: Lens' ListPermissionSetProvisioningStatus (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listPermissionSetProvisioningStatus_filter :: Lens' ListPermissionSetProvisioningStatus (Maybe OperationStatusFilter)

Filters results based on the passed attribute value.

listPermissionSetProvisioningStatus_maxResults :: Lens' ListPermissionSetProvisioningStatus (Maybe Natural)

The maximum number of results to display for the assignment.

listPermissionSetProvisioningStatus_instanceArn :: Lens' ListPermissionSetProvisioningStatus Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

listPermissionSetProvisioningStatusResponse_permissionSetsProvisioningStatus :: Lens' ListPermissionSetProvisioningStatusResponse (Maybe [PermissionSetProvisioningStatusMetadata])

The status object for the permission set provisioning operation.

listPermissionSetProvisioningStatusResponse_nextToken :: Lens' ListPermissionSetProvisioningStatusResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listPermissionSetProvisioningStatusResponse_httpStatus :: Lens' ListPermissionSetProvisioningStatusResponse Int

The response's http status code.

CreateAccountAssignment

createAccountAssignment_instanceArn :: Lens' CreateAccountAssignment Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

createAccountAssignment_targetId :: Lens' CreateAccountAssignment Text

TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

createAccountAssignment_targetType :: Lens' CreateAccountAssignment TargetType

The entity type for which the assignment will be created.

createAccountAssignment_permissionSetArn :: Lens' CreateAccountAssignment Text

The ARN of the permission set that the admin wants to grant the principal access to.

createAccountAssignment_principalType :: Lens' CreateAccountAssignment PrincipalType

The entity type for which the assignment will be created.

createAccountAssignment_principalId :: Lens' CreateAccountAssignment Text

An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

createAccountAssignmentResponse_accountAssignmentCreationStatus :: Lens' CreateAccountAssignmentResponse (Maybe AccountAssignmentOperationStatus)

The status object for the account assignment creation operation.

createAccountAssignmentResponse_httpStatus :: Lens' CreateAccountAssignmentResponse Int

The response's http status code.

Types

AccessControlAttribute

accessControlAttribute_key :: Lens' AccessControlAttribute Text

The name of the attribute associated with your identities in your identity source. This is used to map a specified attribute in your identity source with an attribute in Amazon Web Services SSO.

accessControlAttribute_value :: Lens' AccessControlAttribute AccessControlAttributeValue

The value used for mapping a specified attribute to an identity source.

AccessControlAttributeValue

accessControlAttributeValue_source :: Lens' AccessControlAttributeValue (NonEmpty Text)

The identity source to use when mapping a specified attribute to Amazon Web Services SSO.

AccountAssignment

accountAssignment_principalId :: Lens' AccountAssignment (Maybe Text)

An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

accountAssignment_principalType :: Lens' AccountAssignment (Maybe PrincipalType)

The entity type for which the assignment will be created.

accountAssignment_accountId :: Lens' AccountAssignment (Maybe Text)

The identifier of the Amazon Web Services account.

accountAssignment_permissionSetArn :: Lens' AccountAssignment (Maybe Text)

The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

AccountAssignmentOperationStatus

accountAssignmentOperationStatus_requestId :: Lens' AccountAssignmentOperationStatus (Maybe Text)

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

accountAssignmentOperationStatus_status :: Lens' AccountAssignmentOperationStatus (Maybe StatusValues)

The status of the permission set provisioning process.

accountAssignmentOperationStatus_failureReason :: Lens' AccountAssignmentOperationStatus (Maybe Text)

The message that contains an error or exception in case of an operation failure.

accountAssignmentOperationStatus_targetId :: Lens' AccountAssignmentOperationStatus (Maybe Text)

TargetID is an Amazon Web Services account identifier, typically a 10-12 digit string (For example, 123456789012).

accountAssignmentOperationStatus_principalId :: Lens' AccountAssignmentOperationStatus (Maybe Text)

An identifier for an object in Amazon Web Services SSO, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in Amazon Web Services SSO, see the Amazon Web Services SSO Identity Store API Reference.

accountAssignmentOperationStatus_principalType :: Lens' AccountAssignmentOperationStatus (Maybe PrincipalType)

The entity type for which the assignment will be created.

accountAssignmentOperationStatus_targetType :: Lens' AccountAssignmentOperationStatus (Maybe TargetType)

The entity type for which the assignment will be created.

accountAssignmentOperationStatus_createdDate :: Lens' AccountAssignmentOperationStatus (Maybe UTCTime)

The date that the permission set was created.

accountAssignmentOperationStatus_permissionSetArn :: Lens' AccountAssignmentOperationStatus (Maybe Text)

The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

AccountAssignmentOperationStatusMetadata

accountAssignmentOperationStatusMetadata_requestId :: Lens' AccountAssignmentOperationStatusMetadata (Maybe Text)

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

accountAssignmentOperationStatusMetadata_status :: Lens' AccountAssignmentOperationStatusMetadata (Maybe StatusValues)

The status of the permission set provisioning process.

accountAssignmentOperationStatusMetadata_createdDate :: Lens' AccountAssignmentOperationStatusMetadata (Maybe UTCTime)

The date that the permission set was created.

AttachedManagedPolicy

attachedManagedPolicy_arn :: Lens' AttachedManagedPolicy (Maybe Text)

The ARN of the IAM managed policy. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

attachedManagedPolicy_name :: Lens' AttachedManagedPolicy (Maybe Text)

The name of the IAM managed policy.

InstanceAccessControlAttributeConfiguration

instanceAccessControlAttributeConfiguration_accessControlAttributes :: Lens' InstanceAccessControlAttributeConfiguration [AccessControlAttribute]

Lists the attributes that are configured for ABAC in the specified Amazon Web Services SSO instance.

InstanceMetadata

instanceMetadata_identityStoreId :: Lens' InstanceMetadata (Maybe Text)

The identifier of the identity store that is connected to the SSO instance.

instanceMetadata_instanceArn :: Lens' InstanceMetadata (Maybe Text)

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

OperationStatusFilter

operationStatusFilter_status :: Lens' OperationStatusFilter (Maybe StatusValues)

Filters the list operations result based on the status attribute.

PermissionSet

permissionSet_relayState :: Lens' PermissionSet (Maybe Text)

Used to redirect users within the application during the federation authentication process.

permissionSet_sessionDuration :: Lens' PermissionSet (Maybe Text)

The length of time that the application user sessions are valid for in the ISO-8601 standard.

permissionSet_createdDate :: Lens' PermissionSet (Maybe UTCTime)

The date that the permission set was created.

permissionSet_permissionSetArn :: Lens' PermissionSet (Maybe Text)

The ARN of the permission set. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

permissionSet_name :: Lens' PermissionSet (Maybe Text)

The name of the permission set.

permissionSet_description :: Lens' PermissionSet (Maybe Text)

The description of the PermissionSet.

PermissionSetProvisioningStatus

permissionSetProvisioningStatus_requestId :: Lens' PermissionSetProvisioningStatus (Maybe Text)

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

permissionSetProvisioningStatus_status :: Lens' PermissionSetProvisioningStatus (Maybe StatusValues)

The status of the permission set provisioning process.

permissionSetProvisioningStatus_failureReason :: Lens' PermissionSetProvisioningStatus (Maybe Text)

The message that contains an error or exception in case of an operation failure.

permissionSetProvisioningStatus_accountId :: Lens' PermissionSetProvisioningStatus (Maybe Text)

The identifier of the Amazon Web Services account from which to list the assignments.

permissionSetProvisioningStatus_createdDate :: Lens' PermissionSetProvisioningStatus (Maybe UTCTime)

The date that the permission set was created.

permissionSetProvisioningStatus_permissionSetArn :: Lens' PermissionSetProvisioningStatus (Maybe Text)

The ARN of the permission set that is being provisioned. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

PermissionSetProvisioningStatusMetadata

permissionSetProvisioningStatusMetadata_requestId :: Lens' PermissionSetProvisioningStatusMetadata (Maybe Text)

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

permissionSetProvisioningStatusMetadata_status :: Lens' PermissionSetProvisioningStatusMetadata (Maybe StatusValues)

The status of the permission set provisioning process.

permissionSetProvisioningStatusMetadata_createdDate :: Lens' PermissionSetProvisioningStatusMetadata (Maybe UTCTime)

The date that the permission set was created.

Tag

tag_value :: Lens' Tag (Maybe Text)

The value of the tag.

tag_key :: Lens' Tag (Maybe Text)

The key for the tag.