Amazonka.SSOAdmin.ListAccountsForProvisionedPermissionSet

Description

Lists all the Amazon Web Services accounts where the specified permission set is provisioned.

This operation returns paginated results.

Creating a Request

data ListAccountsForProvisionedPermissionSet

See: newListAccountsForProvisionedPermissionSet smart constructor.

Constructors

ListAccountsForProvisionedPermissionSet'

Fields

provisioningStatus :: Maybe ProvisioningStatus

The permission set provisioning status for an Amazon Web Services account.

nextToken :: Maybe Text

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

maxResults :: Maybe Natural

The maximum number of results to display for the PermissionSet.

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

permissionSetArn :: Text

The ARN of the PermissionSet from which the associated Amazon Web Services accounts will be listed.

newListAccountsForProvisionedPermissionSet

Arguments

:: Text
$sel:instanceArn:ListAccountsForProvisionedPermissionSet'

-> Text
$sel:permissionSetArn:ListAccountsForProvisionedPermissionSet'

-> ListAccountsForProvisionedPermissionSet

Create a value of ListAccountsForProvisionedPermissionSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisioningStatus:ListAccountsForProvisionedPermissionSet', listAccountsForProvisionedPermissionSet_provisioningStatus - The permission set provisioning status for an Amazon Web Services account.

$sel:nextToken:ListAccountsForProvisionedPermissionSet', listAccountsForProvisionedPermissionSet_nextToken - The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

$sel:maxResults:ListAccountsForProvisionedPermissionSet', listAccountsForProvisionedPermissionSet_maxResults - The maximum number of results to display for the PermissionSet.

$sel:instanceArn:ListAccountsForProvisionedPermissionSet', listAccountsForProvisionedPermissionSet_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

$sel:permissionSetArn:ListAccountsForProvisionedPermissionSet', listAccountsForProvisionedPermissionSet_permissionSetArn - The ARN of the PermissionSet from which the associated Amazon Web Services accounts will be listed.

Request Lenses

listAccountsForProvisionedPermissionSet_provisioningStatus :: Lens' ListAccountsForProvisionedPermissionSet (Maybe ProvisioningStatus)

The permission set provisioning status for an Amazon Web Services account.

listAccountsForProvisionedPermissionSet_nextToken :: Lens' ListAccountsForProvisionedPermissionSet (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listAccountsForProvisionedPermissionSet_maxResults :: Lens' ListAccountsForProvisionedPermissionSet (Maybe Natural)

The maximum number of results to display for the PermissionSet.

listAccountsForProvisionedPermissionSet_instanceArn :: Lens' ListAccountsForProvisionedPermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

listAccountsForProvisionedPermissionSet_permissionSetArn :: Lens' ListAccountsForProvisionedPermissionSet Text

The ARN of the PermissionSet from which the associated Amazon Web Services accounts will be listed.

Destructuring the Response

data ListAccountsForProvisionedPermissionSetResponse

See: newListAccountsForProvisionedPermissionSetResponse smart constructor.

Constructors

ListAccountsForProvisionedPermissionSetResponse'

Fields

accountIds :: Maybe [Text]

The list of Amazon Web Services AccountIds.

nextToken :: Maybe Text

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

httpStatus :: Int

The response's http status code.

newListAccountsForProvisionedPermissionSetResponse

Arguments

:: Int
$sel:httpStatus:ListAccountsForProvisionedPermissionSetResponse'

-> ListAccountsForProvisionedPermissionSetResponse

Create a value of ListAccountsForProvisionedPermissionSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIds:ListAccountsForProvisionedPermissionSetResponse', listAccountsForProvisionedPermissionSetResponse_accountIds - The list of Amazon Web Services AccountIds.

$sel:nextToken:ListAccountsForProvisionedPermissionSet', listAccountsForProvisionedPermissionSetResponse_nextToken - The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

$sel:httpStatus:ListAccountsForProvisionedPermissionSetResponse', listAccountsForProvisionedPermissionSetResponse_httpStatus - The response's http status code.

Response Lenses

listAccountsForProvisionedPermissionSetResponse_accountIds :: Lens' ListAccountsForProvisionedPermissionSetResponse (Maybe [Text])

The list of Amazon Web Services AccountIds.

listAccountsForProvisionedPermissionSetResponse_nextToken :: Lens' ListAccountsForProvisionedPermissionSetResponse (Maybe Text)

The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

listAccountsForProvisionedPermissionSetResponse_httpStatus :: Lens' ListAccountsForProvisionedPermissionSetResponse Int

The response's http status code.