Amazonka.SSOAdmin.PutInlinePolicyToPermissionSet

Description

Attaches an IAM inline policy to a permission set.

If the permission set is already referenced by one or more account assignments, you will need to call ProvisionPermissionSet after this action to apply the corresponding IAM policy updates to all assigned accounts.

Creating a Request

data PutInlinePolicyToPermissionSet

See: newPutInlinePolicyToPermissionSet smart constructor.

Constructors

PutInlinePolicyToPermissionSet'

Fields

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

permissionSetArn :: Text

The ARN of the permission set.

inlinePolicy :: Sensitive Text

The IAM inline policy to attach to a PermissionSet.

newPutInlinePolicyToPermissionSet

Arguments

:: Text
$sel:instanceArn:PutInlinePolicyToPermissionSet'

-> Text
$sel:permissionSetArn:PutInlinePolicyToPermissionSet'

-> Text
$sel:inlinePolicy:PutInlinePolicyToPermissionSet'

-> PutInlinePolicyToPermissionSet

Create a value of PutInlinePolicyToPermissionSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceArn:PutInlinePolicyToPermissionSet', putInlinePolicyToPermissionSet_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

$sel:permissionSetArn:PutInlinePolicyToPermissionSet', putInlinePolicyToPermissionSet_permissionSetArn - The ARN of the permission set.

$sel:inlinePolicy:PutInlinePolicyToPermissionSet', putInlinePolicyToPermissionSet_inlinePolicy - The IAM inline policy to attach to a PermissionSet.

Request Lenses

putInlinePolicyToPermissionSet_instanceArn :: Lens' PutInlinePolicyToPermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

putInlinePolicyToPermissionSet_permissionSetArn :: Lens' PutInlinePolicyToPermissionSet Text

The ARN of the permission set.

putInlinePolicyToPermissionSet_inlinePolicy :: Lens' PutInlinePolicyToPermissionSet Text

The IAM inline policy to attach to a PermissionSet.

Destructuring the Response

data PutInlinePolicyToPermissionSetResponse

See: newPutInlinePolicyToPermissionSetResponse smart constructor.

Constructors

PutInlinePolicyToPermissionSetResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutInlinePolicyToPermissionSetResponse

Arguments

:: Int
$sel:httpStatus:PutInlinePolicyToPermissionSetResponse'

-> PutInlinePolicyToPermissionSetResponse

Create a value of PutInlinePolicyToPermissionSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutInlinePolicyToPermissionSetResponse', putInlinePolicyToPermissionSetResponse_httpStatus - The response's http status code.

Response Lenses

putInlinePolicyToPermissionSetResponse_httpStatus :: Lens' PutInlinePolicyToPermissionSetResponse Int

The response's http status code.