Amazonka.SSOAdmin.Types.PermissionSetProvisioningStatus

Description

Documentation

data PermissionSetProvisioningStatus

A structure that is used to provide the status of the provisioning operation for a specified permission set.

See: newPermissionSetProvisioningStatus smart constructor.

Constructors

PermissionSetProvisioningStatus'

Fields

requestId :: Maybe Text

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

status :: Maybe StatusValues

The status of the permission set provisioning process.

failureReason :: Maybe Text

The message that contains an error or exception in case of an operation failure.

accountId :: Maybe Text

The identifier of the Amazon Web Services account from which to list the assignments.

createdDate :: Maybe POSIX

The date that the permission set was created.

permissionSetArn :: Maybe Text

The ARN of the permission set that is being provisioned. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

newPermissionSetProvisioningStatus :: PermissionSetProvisioningStatus

Create a value of PermissionSetProvisioningStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:PermissionSetProvisioningStatus', permissionSetProvisioningStatus_requestId - The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

$sel:status:PermissionSetProvisioningStatus', permissionSetProvisioningStatus_status - The status of the permission set provisioning process.

$sel:failureReason:PermissionSetProvisioningStatus', permissionSetProvisioningStatus_failureReason - The message that contains an error or exception in case of an operation failure.

$sel:accountId:PermissionSetProvisioningStatus', permissionSetProvisioningStatus_accountId - The identifier of the Amazon Web Services account from which to list the assignments.

$sel:createdDate:PermissionSetProvisioningStatus', permissionSetProvisioningStatus_createdDate - The date that the permission set was created.

$sel:permissionSetArn:PermissionSetProvisioningStatus', permissionSetProvisioningStatus_permissionSetArn - The ARN of the permission set that is being provisioned. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

permissionSetProvisioningStatus_requestId :: Lens' PermissionSetProvisioningStatus (Maybe Text)

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

permissionSetProvisioningStatus_status :: Lens' PermissionSetProvisioningStatus (Maybe StatusValues)

The status of the permission set provisioning process.

permissionSetProvisioningStatus_failureReason :: Lens' PermissionSetProvisioningStatus (Maybe Text)

The message that contains an error or exception in case of an operation failure.

permissionSetProvisioningStatus_accountId :: Lens' PermissionSetProvisioningStatus (Maybe Text)

The identifier of the Amazon Web Services account from which to list the assignments.

permissionSetProvisioningStatus_createdDate :: Lens' PermissionSetProvisioningStatus (Maybe UTCTime)

The date that the permission set was created.

permissionSetProvisioningStatus_permissionSetArn :: Lens' PermissionSetProvisioningStatus (Maybe Text)

The ARN of the permission set that is being provisioned. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.