Amazonka.SSOAdmin.Types.PermissionSetProvisioningStatusMetadata

Description

Documentation

data PermissionSetProvisioningStatusMetadata

Provides information about the permission set provisioning status.

See: newPermissionSetProvisioningStatusMetadata smart constructor.

Constructors

PermissionSetProvisioningStatusMetadata'

Fields

requestId :: Maybe Text

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

status :: Maybe StatusValues

The status of the permission set provisioning process.

createdDate :: Maybe POSIX

The date that the permission set was created.

newPermissionSetProvisioningStatusMetadata :: PermissionSetProvisioningStatusMetadata

Create a value of PermissionSetProvisioningStatusMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:PermissionSetProvisioningStatusMetadata', permissionSetProvisioningStatusMetadata_requestId - The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

$sel:status:PermissionSetProvisioningStatusMetadata', permissionSetProvisioningStatusMetadata_status - The status of the permission set provisioning process.

$sel:createdDate:PermissionSetProvisioningStatusMetadata', permissionSetProvisioningStatusMetadata_createdDate - The date that the permission set was created.

permissionSetProvisioningStatusMetadata_requestId :: Lens' PermissionSetProvisioningStatusMetadata (Maybe Text)

The identifier for tracking the request operation that is generated by the universally unique identifier (UUID) workflow.

permissionSetProvisioningStatusMetadata_status :: Lens' PermissionSetProvisioningStatusMetadata (Maybe StatusValues)

The status of the permission set provisioning process.

permissionSetProvisioningStatusMetadata_createdDate :: Lens' PermissionSetProvisioningStatusMetadata (Maybe UTCTime)

The date that the permission set was created.