Amazonka.SSOAdmin.UpdatePermissionSet

Description

Updates an existing permission set.

Creating a Request

data UpdatePermissionSet

See: newUpdatePermissionSet smart constructor.

Constructors

UpdatePermissionSet'

Fields

relayState :: Maybe Text

Used to redirect users within the application during the federation authentication process.

sessionDuration :: Maybe Text

The length of time that the application user sessions are valid for in the ISO-8601 standard.

description :: Maybe Text

The description of the PermissionSet.

instanceArn :: Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

permissionSetArn :: Text

The ARN of the permission set.

newUpdatePermissionSet

Arguments

:: Text
$sel:instanceArn:UpdatePermissionSet'

-> Text
$sel:permissionSetArn:UpdatePermissionSet'

-> UpdatePermissionSet

Create a value of UpdatePermissionSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relayState:UpdatePermissionSet', updatePermissionSet_relayState - Used to redirect users within the application during the federation authentication process.

$sel:sessionDuration:UpdatePermissionSet', updatePermissionSet_sessionDuration - The length of time that the application user sessions are valid for in the ISO-8601 standard.

$sel:description:UpdatePermissionSet', updatePermissionSet_description - The description of the PermissionSet.

$sel:instanceArn:UpdatePermissionSet', updatePermissionSet_instanceArn - The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

$sel:permissionSetArn:UpdatePermissionSet', updatePermissionSet_permissionSetArn - The ARN of the permission set.

Request Lenses

updatePermissionSet_relayState :: Lens' UpdatePermissionSet (Maybe Text)

Used to redirect users within the application during the federation authentication process.

updatePermissionSet_sessionDuration :: Lens' UpdatePermissionSet (Maybe Text)

The length of time that the application user sessions are valid for in the ISO-8601 standard.

updatePermissionSet_description :: Lens' UpdatePermissionSet (Maybe Text)

The description of the PermissionSet.

updatePermissionSet_instanceArn :: Lens' UpdatePermissionSet Text

The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

updatePermissionSet_permissionSetArn :: Lens' UpdatePermissionSet Text

The ARN of the permission set.

Destructuring the Response

data UpdatePermissionSetResponse

See: newUpdatePermissionSetResponse smart constructor.

Constructors

UpdatePermissionSetResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdatePermissionSetResponse

Arguments

:: Int
$sel:httpStatus:UpdatePermissionSetResponse'

-> UpdatePermissionSetResponse

Create a value of UpdatePermissionSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdatePermissionSetResponse', updatePermissionSetResponse_httpStatus - The response's http status code.

Response Lenses

updatePermissionSetResponse_httpStatus :: Lens' UpdatePermissionSetResponse Int

The response's http status code.