Amazonka.SSOOIDC.CreateToken

Description

Creates and returns an access token for the authorized client. The access token issued will be used to fetch short-term credentials for the assigned roles in the AWS account.

Creating a Request

data CreateToken

See: newCreateToken smart constructor.

Constructors

CreateToken'

Fields

redirectUri :: Maybe Text

The location of the application that will receive the authorization code. Users authorize the service to send the request to this location.

refreshToken :: Maybe Text

The token used to obtain an access token in the event that the access token is invalid or expired. This token is not issued by the service.

scope :: Maybe [Text]

The list of scopes that is defined by the client. Upon authorization, this list is used to restrict permissions when granting an access token.

code :: Maybe Text

The authorization code received from the authorization service. This parameter is required to perform an authorization grant request to get access to a token.

clientId :: Text

The unique identifier string for each client. This value should come from the persisted result of the RegisterClient API.

clientSecret :: Text

A secret string generated for the client. This value should come from the persisted result of the RegisterClient API.

grantType :: Text

Supports grant types for authorization code, refresh token, and device code request.

deviceCode :: Text

Used only when calling this API for the device code grant type. This short-term code is used to identify this authentication attempt. This should come from an in-memory reference to the result of the StartDeviceAuthorization API.

newCreateToken

Arguments

:: Text
$sel:clientId:CreateToken'

-> Text
$sel:clientSecret:CreateToken'

-> Text
$sel:grantType:CreateToken'

-> Text
$sel:deviceCode:CreateToken'

-> CreateToken

Create a value of CreateToken with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:redirectUri:CreateToken', createToken_redirectUri - The location of the application that will receive the authorization code. Users authorize the service to send the request to this location.

$sel:refreshToken:CreateToken', createToken_refreshToken - The token used to obtain an access token in the event that the access token is invalid or expired. This token is not issued by the service.

$sel:scope:CreateToken', createToken_scope - The list of scopes that is defined by the client. Upon authorization, this list is used to restrict permissions when granting an access token.

$sel:code:CreateToken', createToken_code - The authorization code received from the authorization service. This parameter is required to perform an authorization grant request to get access to a token.

$sel:clientId:CreateToken', createToken_clientId - The unique identifier string for each client. This value should come from the persisted result of the RegisterClient API.

$sel:clientSecret:CreateToken', createToken_clientSecret - A secret string generated for the client. This value should come from the persisted result of the RegisterClient API.

$sel:grantType:CreateToken', createToken_grantType - Supports grant types for authorization code, refresh token, and device code request.

$sel:deviceCode:CreateToken', createToken_deviceCode - Used only when calling this API for the device code grant type. This short-term code is used to identify this authentication attempt. This should come from an in-memory reference to the result of the StartDeviceAuthorization API.

Request Lenses

createToken_redirectUri :: Lens' CreateToken (Maybe Text)

The location of the application that will receive the authorization code. Users authorize the service to send the request to this location.

createToken_refreshToken :: Lens' CreateToken (Maybe Text)

The token used to obtain an access token in the event that the access token is invalid or expired. This token is not issued by the service.

createToken_scope :: Lens' CreateToken (Maybe [Text])

The list of scopes that is defined by the client. Upon authorization, this list is used to restrict permissions when granting an access token.

createToken_code :: Lens' CreateToken (Maybe Text)

The authorization code received from the authorization service. This parameter is required to perform an authorization grant request to get access to a token.

createToken_clientId :: Lens' CreateToken Text

The unique identifier string for each client. This value should come from the persisted result of the RegisterClient API.

createToken_clientSecret :: Lens' CreateToken Text

A secret string generated for the client. This value should come from the persisted result of the RegisterClient API.

createToken_grantType :: Lens' CreateToken Text

Supports grant types for authorization code, refresh token, and device code request.

createToken_deviceCode :: Lens' CreateToken Text

Used only when calling this API for the device code grant type. This short-term code is used to identify this authentication attempt. This should come from an in-memory reference to the result of the StartDeviceAuthorization API.

Destructuring the Response

data CreateTokenResponse

See: newCreateTokenResponse smart constructor.

Constructors

CreateTokenResponse'

Fields

accessToken :: Maybe Text

An opaque token to access AWS SSO resources assigned to a user.

refreshToken :: Maybe Text

A token that, if present, can be used to refresh a previously issued access token that might have expired.

expiresIn :: Maybe Int

Indicates the time in seconds when an access token will expire.

tokenType :: Maybe Text

Used to notify the client that the returned token is an access token. The supported type is BearerToken.

idToken :: Maybe Text

The identifier of the user that associated with the access token, if present.

httpStatus :: Int

The response's http status code.

newCreateTokenResponse

Arguments

:: Int
$sel:httpStatus:CreateTokenResponse'

-> CreateTokenResponse

Create a value of CreateTokenResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessToken:CreateTokenResponse', createTokenResponse_accessToken - An opaque token to access AWS SSO resources assigned to a user.

$sel:refreshToken:CreateToken', createTokenResponse_refreshToken - A token that, if present, can be used to refresh a previously issued access token that might have expired.

$sel:expiresIn:CreateTokenResponse', createTokenResponse_expiresIn - Indicates the time in seconds when an access token will expire.

$sel:tokenType:CreateTokenResponse', createTokenResponse_tokenType - Used to notify the client that the returned token is an access token. The supported type is BearerToken.

$sel:idToken:CreateTokenResponse', createTokenResponse_idToken - The identifier of the user that associated with the access token, if present.

$sel:httpStatus:CreateTokenResponse', createTokenResponse_httpStatus - The response's http status code.

Response Lenses

createTokenResponse_accessToken :: Lens' CreateTokenResponse (Maybe Text)

An opaque token to access AWS SSO resources assigned to a user.

createTokenResponse_refreshToken :: Lens' CreateTokenResponse (Maybe Text)

A token that, if present, can be used to refresh a previously issued access token that might have expired.

createTokenResponse_expiresIn :: Lens' CreateTokenResponse (Maybe Int)

Indicates the time in seconds when an access token will expire.

createTokenResponse_tokenType :: Lens' CreateTokenResponse (Maybe Text)

Used to notify the client that the returned token is an access token. The supported type is BearerToken.

createTokenResponse_idToken :: Lens' CreateTokenResponse (Maybe Text)

The identifier of the user that associated with the access token, if present.

createTokenResponse_httpStatus :: Lens' CreateTokenResponse Int

The response's http status code.