Amazonka.SSOOIDC.Lens

Description

Operations

StartDeviceAuthorization

startDeviceAuthorization_clientId :: Lens' StartDeviceAuthorization Text

The unique identifier string for the client that is registered with AWS SSO. This value should come from the persisted result of the RegisterClient API operation.

startDeviceAuthorization_clientSecret :: Lens' StartDeviceAuthorization Text

A secret string that is generated for the client. This value should come from the persisted result of the RegisterClient API operation.

startDeviceAuthorization_startUrl :: Lens' StartDeviceAuthorization Text

The URL for the AWS SSO user portal. For more information, see Using the User Portal in the AWS Single Sign-On User Guide.

startDeviceAuthorizationResponse_userCode :: Lens' StartDeviceAuthorizationResponse (Maybe Text)

A one-time user verification code. This is needed to authorize an in-use device.

startDeviceAuthorizationResponse_interval :: Lens' StartDeviceAuthorizationResponse (Maybe Int)

Indicates the number of seconds the client must wait between attempts when polling for a session.

startDeviceAuthorizationResponse_expiresIn :: Lens' StartDeviceAuthorizationResponse (Maybe Int)

Indicates the number of seconds in which the verification code will become invalid.

startDeviceAuthorizationResponse_verificationUri :: Lens' StartDeviceAuthorizationResponse (Maybe Text)

The URI of the verification page that takes the userCode to authorize the device.

startDeviceAuthorizationResponse_deviceCode :: Lens' StartDeviceAuthorizationResponse (Maybe Text)

The short-lived code that is used by the device when polling for a session token.

startDeviceAuthorizationResponse_verificationUriComplete :: Lens' StartDeviceAuthorizationResponse (Maybe Text)

An alternate URL that the client can use to automatically launch a browser. This process skips the manual step in which the user visits the verification page and enters their code.

startDeviceAuthorizationResponse_httpStatus :: Lens' StartDeviceAuthorizationResponse Int

The response's http status code.

CreateToken

createToken_redirectUri :: Lens' CreateToken (Maybe Text)

The location of the application that will receive the authorization code. Users authorize the service to send the request to this location.

createToken_refreshToken :: Lens' CreateToken (Maybe Text)

The token used to obtain an access token in the event that the access token is invalid or expired. This token is not issued by the service.

createToken_scope :: Lens' CreateToken (Maybe [Text])

The list of scopes that is defined by the client. Upon authorization, this list is used to restrict permissions when granting an access token.

createToken_code :: Lens' CreateToken (Maybe Text)

The authorization code received from the authorization service. This parameter is required to perform an authorization grant request to get access to a token.

createToken_clientId :: Lens' CreateToken Text

The unique identifier string for each client. This value should come from the persisted result of the RegisterClient API.

createToken_clientSecret :: Lens' CreateToken Text

A secret string generated for the client. This value should come from the persisted result of the RegisterClient API.

createToken_grantType :: Lens' CreateToken Text

Supports grant types for authorization code, refresh token, and device code request.

createToken_deviceCode :: Lens' CreateToken Text

Used only when calling this API for the device code grant type. This short-term code is used to identify this authentication attempt. This should come from an in-memory reference to the result of the StartDeviceAuthorization API.

createTokenResponse_accessToken :: Lens' CreateTokenResponse (Maybe Text)

An opaque token to access AWS SSO resources assigned to a user.

createTokenResponse_refreshToken :: Lens' CreateTokenResponse (Maybe Text)

A token that, if present, can be used to refresh a previously issued access token that might have expired.

createTokenResponse_expiresIn :: Lens' CreateTokenResponse (Maybe Int)

Indicates the time in seconds when an access token will expire.

createTokenResponse_tokenType :: Lens' CreateTokenResponse (Maybe Text)

Used to notify the client that the returned token is an access token. The supported type is BearerToken.

createTokenResponse_idToken :: Lens' CreateTokenResponse (Maybe Text)

The identifier of the user that associated with the access token, if present.

createTokenResponse_httpStatus :: Lens' CreateTokenResponse Int

The response's http status code.

RegisterClient

registerClient_scopes :: Lens' RegisterClient (Maybe [Text])

The list of scopes that are defined by the client. Upon authorization, this list is used to restrict permissions when granting an access token.

registerClient_clientName :: Lens' RegisterClient Text

The friendly name of the client.

registerClient_clientType :: Lens' RegisterClient Text

The type of client. The service supports only public as a client type. Anything other than public will be rejected by the service.

registerClientResponse_clientId :: Lens' RegisterClientResponse (Maybe Text)

The unique identifier string for each client. This client uses this identifier to get authenticated by the service in subsequent calls.

registerClientResponse_clientSecret :: Lens' RegisterClientResponse (Maybe Text)

A secret string generated for the client. The client will use this string to get authenticated by the service in subsequent calls.

registerClientResponse_clientIdIssuedAt :: Lens' RegisterClientResponse (Maybe Integer)

Indicates the time at which the clientId and clientSecret were issued.

registerClientResponse_clientSecretExpiresAt :: Lens' RegisterClientResponse (Maybe Integer)

Indicates the time at which the clientId and clientSecret will become invalid.

registerClientResponse_tokenEndpoint :: Lens' RegisterClientResponse (Maybe Text)

The endpoint where the client can get an access token.

registerClientResponse_authorizationEndpoint :: Lens' RegisterClientResponse (Maybe Text)

The endpoint where the client can request authorization.

registerClientResponse_httpStatus :: Lens' RegisterClientResponse Int

The response's http status code.

Types