Amazonka.SSOOIDC.Types

Description

Service Configuration

defaultService :: Service

API version 2019-06-10 of the Amazon SSO OIDC SDK configuration.

Errors

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that something is wrong with the input to the request. For example, a required parameter might be missing or out of range.

_InvalidScopeException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the scope provided in the request is invalid.

_UnsupportedGrantTypeException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the grant type in the request is not supported by the service.

_InvalidClientException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the clientId or clientSecret in the request is invalid. For example, this can occur when a client sends an incorrect clientId or an expired clientSecret.

_InvalidClientMetadataException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the client information sent in the request during registration is invalid.

_UnauthorizedClientException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the client is not currently authorized to make the request. This can happen when a clientId is not issued for a public client.

_SlowDownException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the client is making the request too frequently and is more than the service can handle.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that an error from the service occurred while trying to process a request.

_InvalidGrantException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a request contains an invalid grant. This can occur if a client makes a CreateToken request with an invalid grant type.

_AuthorizationPendingException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a request to authorize a client with an access user session token is pending.

_ExpiredTokenException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the token issued by the service is expired and is no longer valid.