Amazonka.SWF.CountClosedWorkflowExecutions

Description

Returns the number of closed workflow executions within the given domain that meet the specified filtering criteria.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tagFilter.tag: String constraint. The key is swf:tagFilter.tag.

typeFilter.name: String constraint. The key is swf:typeFilter.name.

typeFilter.version: String constraint. The key is swf:typeFilter.version.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

Creating a Request

data CountClosedWorkflowExecutions

See: newCountClosedWorkflowExecutions smart constructor.

Constructors

CountClosedWorkflowExecutions'

Fields

executionFilter :: Maybe WorkflowExecutionFilter

If specified, only workflow executions matching the WorkflowId in the filter are counted.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

closeStatusFilter :: Maybe CloseStatusFilter

If specified, only workflow executions that match this close status are counted. This filter has an affect only if executionStatus is specified as CLOSED.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

typeFilter :: Maybe WorkflowTypeFilter

If specified, indicates the type of the workflow executions to be counted.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

closeTimeFilter :: Maybe ExecutionTimeFilter

If specified, only workflow executions that meet the close time criteria of the filter are counted.

startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

tagFilter :: Maybe TagFilter

If specified, only executions that have a tag that matches the filter are counted.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

startTimeFilter :: Maybe ExecutionTimeFilter

If specified, only workflow executions that meet the start time criteria of the filter are counted.

startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

domain :: Text

The name of the domain containing the workflow executions to count.

newCountClosedWorkflowExecutions

Arguments

:: Text
$sel:domain:CountClosedWorkflowExecutions'

-> CountClosedWorkflowExecutions

Create a value of CountClosedWorkflowExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionFilter:CountClosedWorkflowExecutions', countClosedWorkflowExecutions_executionFilter - If specified, only workflow executions matching the WorkflowId in the filter are counted.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:closeStatusFilter:CountClosedWorkflowExecutions', countClosedWorkflowExecutions_closeStatusFilter - If specified, only workflow executions that match this close status are counted. This filter has an affect only if executionStatus is specified as CLOSED.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:typeFilter:CountClosedWorkflowExecutions', countClosedWorkflowExecutions_typeFilter - If specified, indicates the type of the workflow executions to be counted.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:closeTimeFilter:CountClosedWorkflowExecutions', countClosedWorkflowExecutions_closeTimeFilter - If specified, only workflow executions that meet the close time criteria of the filter are counted.

startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

$sel:tagFilter:CountClosedWorkflowExecutions', countClosedWorkflowExecutions_tagFilter - If specified, only executions that have a tag that matches the filter are counted.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:startTimeFilter:CountClosedWorkflowExecutions', countClosedWorkflowExecutions_startTimeFilter - If specified, only workflow executions that meet the start time criteria of the filter are counted.

startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

$sel:domain:CountClosedWorkflowExecutions', countClosedWorkflowExecutions_domain - The name of the domain containing the workflow executions to count.

Request Lenses

countClosedWorkflowExecutions_executionFilter :: Lens' CountClosedWorkflowExecutions (Maybe WorkflowExecutionFilter)

If specified, only workflow executions matching the WorkflowId in the filter are counted.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

countClosedWorkflowExecutions_closeStatusFilter :: Lens' CountClosedWorkflowExecutions (Maybe CloseStatusFilter)

If specified, only workflow executions that match this close status are counted. This filter has an affect only if executionStatus is specified as CLOSED.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

countClosedWorkflowExecutions_typeFilter :: Lens' CountClosedWorkflowExecutions (Maybe WorkflowTypeFilter)

If specified, indicates the type of the workflow executions to be counted.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

countClosedWorkflowExecutions_closeTimeFilter :: Lens' CountClosedWorkflowExecutions (Maybe ExecutionTimeFilter)

If specified, only workflow executions that meet the close time criteria of the filter are counted.

startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

countClosedWorkflowExecutions_tagFilter :: Lens' CountClosedWorkflowExecutions (Maybe TagFilter)

If specified, only executions that have a tag that matches the filter are counted.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

countClosedWorkflowExecutions_startTimeFilter :: Lens' CountClosedWorkflowExecutions (Maybe ExecutionTimeFilter)

If specified, only workflow executions that meet the start time criteria of the filter are counted.

startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

countClosedWorkflowExecutions_domain :: Lens' CountClosedWorkflowExecutions Text

The name of the domain containing the workflow executions to count.

Destructuring the Response

data WorkflowExecutionCount

Contains the count of workflow executions returned from CountOpenWorkflowExecutions or CountClosedWorkflowExecutions

See: newWorkflowExecutionCount smart constructor.

Constructors

WorkflowExecutionCount'

Fields

truncated :: Maybe Bool

If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

count :: Natural

The number of workflow executions.

newWorkflowExecutionCount

Arguments

:: Natural
$sel:count:WorkflowExecutionCount'

-> WorkflowExecutionCount

Create a value of WorkflowExecutionCount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:truncated:WorkflowExecutionCount', workflowExecutionCount_truncated - If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

$sel:count:WorkflowExecutionCount', workflowExecutionCount_count - The number of workflow executions.

Response Lenses

workflowExecutionCount_truncated :: Lens' WorkflowExecutionCount (Maybe Bool)

If set to true, indicates that the actual count was more than the maximum supported by this API and the count returned is the truncated value.

workflowExecutionCount_count :: Lens' WorkflowExecutionCount Natural

The number of workflow executions.