Amazonka.SWF.ListClosedWorkflowExecutions

Description

Returns a list of closed workflow executions in the specified domain that meet the filtering criteria. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

Constrain the following parameters by using a Condition element with the appropriate keys.

tagFilter.tag: String constraint. The key is swf:tagFilter.tag.

typeFilter.name: String constraint. The key is swf:typeFilter.name.

typeFilter.version: String constraint. The key is swf:typeFilter.version.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

This operation returns paginated results.

Creating a Request

data ListClosedWorkflowExecutions

See: newListClosedWorkflowExecutions smart constructor.

Constructors

ListClosedWorkflowExecutions'

Fields

nextPageToken :: Maybe Text

If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

executionFilter :: Maybe WorkflowExecutionFilter

If specified, only workflow executions matching the workflow ID specified in the filter are returned.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

closeStatusFilter :: Maybe CloseStatusFilter

If specified, only workflow executions that match this close status are listed. For example, if TERMINATED is specified, then only TERMINATED workflow executions are listed.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

typeFilter :: Maybe WorkflowTypeFilter

If specified, only executions of the type specified in the filter are returned.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

closeTimeFilter :: Maybe ExecutionTimeFilter

If specified, the workflow executions are included in the returned results based on whether their close times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their close times.

startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

reverseOrder :: Maybe Bool

When set to true, returns the results in reverse order. By default the results are returned in descending order of the start or the close time of the executions.

tagFilter :: Maybe TagFilter

If specified, only executions that have the matching tag are listed.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

startTimeFilter :: Maybe ExecutionTimeFilter

If specified, the workflow executions are included in the returned results based on whether their start times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their start times.

startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

maximumPageSize :: Maybe Natural

The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

domain :: Text

The name of the domain that contains the workflow executions to list.

newListClosedWorkflowExecutions

Arguments

:: Text
$sel:domain:ListClosedWorkflowExecutions'

-> ListClosedWorkflowExecutions

Create a value of ListClosedWorkflowExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:ListClosedWorkflowExecutions', listClosedWorkflowExecutions_nextPageToken - If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

$sel:executionFilter:ListClosedWorkflowExecutions', listClosedWorkflowExecutions_executionFilter - If specified, only workflow executions matching the workflow ID specified in the filter are returned.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:closeStatusFilter:ListClosedWorkflowExecutions', listClosedWorkflowExecutions_closeStatusFilter - If specified, only workflow executions that match this close status are listed. For example, if TERMINATED is specified, then only TERMINATED workflow executions are listed.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:typeFilter:ListClosedWorkflowExecutions', listClosedWorkflowExecutions_typeFilter - If specified, only executions of the type specified in the filter are returned.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:closeTimeFilter:ListClosedWorkflowExecutions', listClosedWorkflowExecutions_closeTimeFilter - If specified, the workflow executions are included in the returned results based on whether their close times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their close times.

startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

$sel:reverseOrder:ListClosedWorkflowExecutions', listClosedWorkflowExecutions_reverseOrder - When set to true, returns the results in reverse order. By default the results are returned in descending order of the start or the close time of the executions.

$sel:tagFilter:ListClosedWorkflowExecutions', listClosedWorkflowExecutions_tagFilter - If specified, only executions that have the matching tag are listed.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

$sel:startTimeFilter:ListClosedWorkflowExecutions', listClosedWorkflowExecutions_startTimeFilter - If specified, the workflow executions are included in the returned results based on whether their start times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their start times.

startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

$sel:maximumPageSize:ListClosedWorkflowExecutions', listClosedWorkflowExecutions_maximumPageSize - The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

$sel:domain:ListClosedWorkflowExecutions', listClosedWorkflowExecutions_domain - The name of the domain that contains the workflow executions to list.

Request Lenses

listClosedWorkflowExecutions_nextPageToken :: Lens' ListClosedWorkflowExecutions (Maybe Text)

If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

listClosedWorkflowExecutions_executionFilter :: Lens' ListClosedWorkflowExecutions (Maybe WorkflowExecutionFilter)

If specified, only workflow executions matching the workflow ID specified in the filter are returned.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

listClosedWorkflowExecutions_closeStatusFilter :: Lens' ListClosedWorkflowExecutions (Maybe CloseStatusFilter)

If specified, only workflow executions that match this close status are listed. For example, if TERMINATED is specified, then only TERMINATED workflow executions are listed.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

listClosedWorkflowExecutions_typeFilter :: Lens' ListClosedWorkflowExecutions (Maybe WorkflowTypeFilter)

If specified, only executions of the type specified in the filter are returned.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

listClosedWorkflowExecutions_closeTimeFilter :: Lens' ListClosedWorkflowExecutions (Maybe ExecutionTimeFilter)

If specified, the workflow executions are included in the returned results based on whether their close times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their close times.

startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

listClosedWorkflowExecutions_reverseOrder :: Lens' ListClosedWorkflowExecutions (Maybe Bool)

When set to true, returns the results in reverse order. By default the results are returned in descending order of the start or the close time of the executions.

listClosedWorkflowExecutions_tagFilter :: Lens' ListClosedWorkflowExecutions (Maybe TagFilter)

If specified, only executions that have the matching tag are listed.

closeStatusFilter, executionFilter, typeFilter and tagFilter are mutually exclusive. You can specify at most one of these in a request.

listClosedWorkflowExecutions_startTimeFilter :: Lens' ListClosedWorkflowExecutions (Maybe ExecutionTimeFilter)

If specified, the workflow executions are included in the returned results based on whether their start times are within the range specified by this filter. Also, if this parameter is specified, the returned results are ordered by their start times.

startTimeFilter and closeTimeFilter are mutually exclusive. You must specify one of these in a request but not both.

listClosedWorkflowExecutions_maximumPageSize :: Lens' ListClosedWorkflowExecutions (Maybe Natural)

The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

listClosedWorkflowExecutions_domain :: Lens' ListClosedWorkflowExecutions Text

The name of the domain that contains the workflow executions to list.

Destructuring the Response

data WorkflowExecutionInfos

Contains a paginated list of information about workflow executions.

See: newWorkflowExecutionInfos smart constructor.

Constructors

WorkflowExecutionInfos'

Fields

nextPageToken :: Maybe Text

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

executionInfos :: [WorkflowExecutionInfo]

The list of workflow information structures.

newWorkflowExecutionInfos :: WorkflowExecutionInfos

Create a value of WorkflowExecutionInfos with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:WorkflowExecutionInfos', workflowExecutionInfos_nextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

$sel:executionInfos:WorkflowExecutionInfos', workflowExecutionInfos_executionInfos - The list of workflow information structures.

Response Lenses

workflowExecutionInfos_nextPageToken :: Lens' WorkflowExecutionInfos (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

workflowExecutionInfos_executionInfos :: Lens' WorkflowExecutionInfos [WorkflowExecutionInfo]

The list of workflow information structures.