Amazonka.SWF.ListDomains

Description

Returns the list of domains registered in the account. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextPageToken returned by the initial call.

This operation is eventually consistent. The results are best effort and may not exactly reflect recent updates and changes.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains. The element must be set to arn:aws:swf::AccountID:domain/*, where AccountID is the account ID, with no dashes.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

This operation returns paginated results.

Creating a Request

data ListDomains

See: newListDomains smart constructor.

Constructors

ListDomains'

Fields

nextPageToken :: Maybe Text

If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

reverseOrder :: Maybe Bool

When set to true, returns the results in reverse order. By default, the results are returned in ascending alphabetical order by name of the domains.

maximumPageSize :: Maybe Natural

The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

registrationStatus :: RegistrationStatus

Specifies the registration status of the domains to list.

newListDomains

Arguments

:: RegistrationStatus
$sel:registrationStatus:ListDomains'

-> ListDomains

Create a value of ListDomains with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:ListDomains', listDomains_nextPageToken - If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

$sel:reverseOrder:ListDomains', listDomains_reverseOrder - When set to true, returns the results in reverse order. By default, the results are returned in ascending alphabetical order by name of the domains.

$sel:maximumPageSize:ListDomains', listDomains_maximumPageSize - The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

$sel:registrationStatus:ListDomains', listDomains_registrationStatus - Specifies the registration status of the domains to list.

Request Lenses

listDomains_nextPageToken :: Lens' ListDomains (Maybe Text)

If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

listDomains_reverseOrder :: Lens' ListDomains (Maybe Bool)

When set to true, returns the results in reverse order. By default, the results are returned in ascending alphabetical order by name of the domains.

listDomains_maximumPageSize :: Lens' ListDomains (Maybe Natural)

The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

listDomains_registrationStatus :: Lens' ListDomains RegistrationStatus

Specifies the registration status of the domains to list.

Destructuring the Response

data ListDomainsResponse

Contains a paginated collection of DomainInfo structures.

See: newListDomainsResponse smart constructor.

Constructors

ListDomainsResponse'

Fields

nextPageToken :: Maybe Text

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

httpStatus :: Int

The response's http status code.

domainInfos :: [DomainInfo]

A list of DomainInfo structures.

newListDomainsResponse

Arguments

:: Int
$sel:httpStatus:ListDomainsResponse'

-> ListDomainsResponse

Create a value of ListDomainsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:ListDomains', listDomainsResponse_nextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

$sel:httpStatus:ListDomainsResponse', listDomainsResponse_httpStatus - The response's http status code.

$sel:domainInfos:ListDomainsResponse', listDomainsResponse_domainInfos - A list of DomainInfo structures.

Response Lenses

listDomainsResponse_nextPageToken :: Lens' ListDomainsResponse (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

listDomainsResponse_httpStatus :: Lens' ListDomainsResponse Int

The response's http status code.

listDomainsResponse_domainInfos :: Lens' ListDomainsResponse [DomainInfo]

A list of DomainInfo structures.