Amazonka.SWF.ListWorkflowTypes

Description

Returns information about workflow types in the specified domain. The results may be split into multiple pages that can be retrieved by making the call repeatedly.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

This operation returns paginated results.

Creating a Request

data ListWorkflowTypes

See: newListWorkflowTypes smart constructor.

Constructors

ListWorkflowTypes'

Fields

nextPageToken :: Maybe Text

If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

reverseOrder :: Maybe Bool

When set to true, returns the results in reverse order. By default the results are returned in ascending alphabetical order of the name of the workflow types.

name :: Maybe Text

If specified, lists the workflow type with this name.

maximumPageSize :: Maybe Natural

The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

domain :: Text

The name of the domain in which the workflow types have been registered.

registrationStatus :: RegistrationStatus

Specifies the registration status of the workflow types to list.

newListWorkflowTypes

Arguments

:: Text
$sel:domain:ListWorkflowTypes'

-> RegistrationStatus
$sel:registrationStatus:ListWorkflowTypes'

-> ListWorkflowTypes

Create a value of ListWorkflowTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:ListWorkflowTypes', listWorkflowTypes_nextPageToken - If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

$sel:reverseOrder:ListWorkflowTypes', listWorkflowTypes_reverseOrder - When set to true, returns the results in reverse order. By default the results are returned in ascending alphabetical order of the name of the workflow types.

$sel:name:ListWorkflowTypes', listWorkflowTypes_name - If specified, lists the workflow type with this name.

$sel:maximumPageSize:ListWorkflowTypes', listWorkflowTypes_maximumPageSize - The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

$sel:domain:ListWorkflowTypes', listWorkflowTypes_domain - The name of the domain in which the workflow types have been registered.

$sel:registrationStatus:ListWorkflowTypes', listWorkflowTypes_registrationStatus - Specifies the registration status of the workflow types to list.

Request Lenses

listWorkflowTypes_nextPageToken :: Lens' ListWorkflowTypes (Maybe Text)

If NextPageToken is returned there are more results available. The value of NextPageToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 60 seconds. Using an expired pagination token will return a 400 error: "Specified token has exceeded its maximum lifetime".

The configured maximumPageSize determines how many results can be returned in a single call.

listWorkflowTypes_reverseOrder :: Lens' ListWorkflowTypes (Maybe Bool)

When set to true, returns the results in reverse order. By default the results are returned in ascending alphabetical order of the name of the workflow types.

listWorkflowTypes_name :: Lens' ListWorkflowTypes (Maybe Text)

If specified, lists the workflow type with this name.

listWorkflowTypes_maximumPageSize :: Lens' ListWorkflowTypes (Maybe Natural)

The maximum number of results that are returned per call. Use nextPageToken to obtain further pages of results.

listWorkflowTypes_domain :: Lens' ListWorkflowTypes Text

The name of the domain in which the workflow types have been registered.

listWorkflowTypes_registrationStatus :: Lens' ListWorkflowTypes RegistrationStatus

Specifies the registration status of the workflow types to list.

Destructuring the Response

data ListWorkflowTypesResponse

Contains a paginated list of information structures about workflow types.

See: newListWorkflowTypesResponse smart constructor.

Constructors

ListWorkflowTypesResponse'

Fields

nextPageToken :: Maybe Text

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

httpStatus :: Int

The response's http status code.

typeInfos :: [WorkflowTypeInfo]

The list of workflow type information.

newListWorkflowTypesResponse

Arguments

:: Int
$sel:httpStatus:ListWorkflowTypesResponse'

-> ListWorkflowTypesResponse

Create a value of ListWorkflowTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:ListWorkflowTypes', listWorkflowTypesResponse_nextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

$sel:httpStatus:ListWorkflowTypesResponse', listWorkflowTypesResponse_httpStatus - The response's http status code.

$sel:typeInfos:ListWorkflowTypesResponse', listWorkflowTypesResponse_typeInfos - The list of workflow type information.

Response Lenses

listWorkflowTypesResponse_nextPageToken :: Lens' ListWorkflowTypesResponse (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

listWorkflowTypesResponse_httpStatus :: Lens' ListWorkflowTypesResponse Int

The response's http status code.

listWorkflowTypesResponse_typeInfos :: Lens' ListWorkflowTypesResponse [WorkflowTypeInfo]

The list of workflow type information.