Amazonka.SWF.SignalWorkflowExecution

Description

Records a WorkflowExecutionSignaled event in the workflow execution history and creates a decision task for the workflow execution identified by the given domain, workflowId and runId. The event is recorded with the specified user defined signalName and input (if provided).

If a runId isn't specified, then the WorkflowExecutionSignaled event is recorded in the history of the current open workflow with the matching workflowId in the domain.

If the specified workflow execution isn't open, this method fails with UnknownResource.

Access Control

You can use IAM policies to control this action's access to Amazon SWF resources as follows:

Use a Resource element with the domain name to limit the action to only specified domains.

Use an Action element to allow or deny permission to call this action.

You cannot use an IAM policy to constrain this action's parameters.

If the caller doesn't have sufficient permissions to invoke the action, or the parameter values fall outside the specified constraints, the action fails. The associated event attribute's cause parameter is set to OPERATION_NOT_PERMITTED. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

Creating a Request

data SignalWorkflowExecution

See: newSignalWorkflowExecution smart constructor.

Constructors

SignalWorkflowExecution'

Fields

input :: Maybe Text

Data to attach to the WorkflowExecutionSignaled event in the target workflow execution's history.

runId :: Maybe Text

The runId of the workflow execution to signal.

domain :: Text

The name of the domain containing the workflow execution to signal.

workflowId :: Text

The workflowId of the workflow execution to signal.

signalName :: Text

The name of the signal. This name must be meaningful to the target workflow.

newSignalWorkflowExecution

Arguments

:: Text
$sel:domain:SignalWorkflowExecution'

-> Text
$sel:workflowId:SignalWorkflowExecution'

-> Text
$sel:signalName:SignalWorkflowExecution'

-> SignalWorkflowExecution

Create a value of SignalWorkflowExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:input:SignalWorkflowExecution', signalWorkflowExecution_input - Data to attach to the WorkflowExecutionSignaled event in the target workflow execution's history.

$sel:runId:SignalWorkflowExecution', signalWorkflowExecution_runId - The runId of the workflow execution to signal.

$sel:domain:SignalWorkflowExecution', signalWorkflowExecution_domain - The name of the domain containing the workflow execution to signal.

$sel:workflowId:SignalWorkflowExecution', signalWorkflowExecution_workflowId - The workflowId of the workflow execution to signal.

$sel:signalName:SignalWorkflowExecution', signalWorkflowExecution_signalName - The name of the signal. This name must be meaningful to the target workflow.

Request Lenses

signalWorkflowExecution_input :: Lens' SignalWorkflowExecution (Maybe Text)

Data to attach to the WorkflowExecutionSignaled event in the target workflow execution's history.

signalWorkflowExecution_runId :: Lens' SignalWorkflowExecution (Maybe Text)

The runId of the workflow execution to signal.

signalWorkflowExecution_domain :: Lens' SignalWorkflowExecution Text

The name of the domain containing the workflow execution to signal.

signalWorkflowExecution_workflowId :: Lens' SignalWorkflowExecution Text

The workflowId of the workflow execution to signal.

signalWorkflowExecution_signalName :: Lens' SignalWorkflowExecution Text

The name of the signal. This name must be meaningful to the target workflow.

Destructuring the Response

data SignalWorkflowExecutionResponse

See: newSignalWorkflowExecutionResponse smart constructor.

Constructors

SignalWorkflowExecutionResponse'

newSignalWorkflowExecutionResponse :: SignalWorkflowExecutionResponse

Create a value of SignalWorkflowExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.