Amazonka.SWF.Types.ActivityTaskCanceledEventAttributes

Description

Documentation

data ActivityTaskCanceledEventAttributes

Provides the details of the ActivityTaskCanceled event.

See: newActivityTaskCanceledEventAttributes smart constructor.

Constructors

ActivityTaskCanceledEventAttributes'

Fields

latestCancelRequestedEventId :: Maybe Integer

If set, contains the ID of the last ActivityTaskCancelRequested event recorded for this activity task. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

details :: Maybe Text

Details of the cancellation.

scheduledEventId :: Integer

The ID of the ActivityTaskScheduled event that was recorded when this activity task was scheduled. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

startedEventId :: Integer

The ID of the ActivityTaskStarted event recorded when this activity task was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newActivityTaskCanceledEventAttributes

Arguments

:: Integer
$sel:scheduledEventId:ActivityTaskCanceledEventAttributes'

-> Integer
$sel:startedEventId:ActivityTaskCanceledEventAttributes'

-> ActivityTaskCanceledEventAttributes

Create a value of ActivityTaskCanceledEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestCancelRequestedEventId:ActivityTaskCanceledEventAttributes', activityTaskCanceledEventAttributes_latestCancelRequestedEventId - If set, contains the ID of the last ActivityTaskCancelRequested event recorded for this activity task. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

$sel:details:ActivityTaskCanceledEventAttributes', activityTaskCanceledEventAttributes_details - Details of the cancellation.

$sel:scheduledEventId:ActivityTaskCanceledEventAttributes', activityTaskCanceledEventAttributes_scheduledEventId - The ID of the ActivityTaskScheduled event that was recorded when this activity task was scheduled. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

$sel:startedEventId:ActivityTaskCanceledEventAttributes', activityTaskCanceledEventAttributes_startedEventId - The ID of the ActivityTaskStarted event recorded when this activity task was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

activityTaskCanceledEventAttributes_latestCancelRequestedEventId :: Lens' ActivityTaskCanceledEventAttributes (Maybe Integer)

If set, contains the ID of the last ActivityTaskCancelRequested event recorded for this activity task. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

activityTaskCanceledEventAttributes_details :: Lens' ActivityTaskCanceledEventAttributes (Maybe Text)

Details of the cancellation.

activityTaskCanceledEventAttributes_scheduledEventId :: Lens' ActivityTaskCanceledEventAttributes Integer

The ID of the ActivityTaskScheduled event that was recorded when this activity task was scheduled. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

activityTaskCanceledEventAttributes_startedEventId :: Lens' ActivityTaskCanceledEventAttributes Integer

The ID of the ActivityTaskStarted event recorded when this activity task was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.