Amazonka.SWF.Types.ChildWorkflowExecutionCanceledEventAttributes

Description

Documentation

data ChildWorkflowExecutionCanceledEventAttributes

Provide details of the ChildWorkflowExecutionCanceled event.

See: newChildWorkflowExecutionCanceledEventAttributes smart constructor.

Constructors

ChildWorkflowExecutionCanceledEventAttributes'

Fields

details :: Maybe Text

Details of the cancellation (if provided).

workflowExecution :: WorkflowExecution

The child workflow execution that was canceled.

workflowType :: WorkflowType

The type of the child workflow execution.

initiatedEventId :: Integer

The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution Decision to start this child workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

startedEventId :: Integer

The ID of the ChildWorkflowExecutionStarted event recorded when this child workflow execution was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newChildWorkflowExecutionCanceledEventAttributes

Arguments

:: WorkflowExecution
$sel:workflowExecution:ChildWorkflowExecutionCanceledEventAttributes'

-> WorkflowType
$sel:workflowType:ChildWorkflowExecutionCanceledEventAttributes'

-> Integer
$sel:initiatedEventId:ChildWorkflowExecutionCanceledEventAttributes'

-> Integer
$sel:startedEventId:ChildWorkflowExecutionCanceledEventAttributes'

-> ChildWorkflowExecutionCanceledEventAttributes

Create a value of ChildWorkflowExecutionCanceledEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:details:ChildWorkflowExecutionCanceledEventAttributes', childWorkflowExecutionCanceledEventAttributes_details - Details of the cancellation (if provided).

$sel:workflowExecution:ChildWorkflowExecutionCanceledEventAttributes', childWorkflowExecutionCanceledEventAttributes_workflowExecution - The child workflow execution that was canceled.

$sel:workflowType:ChildWorkflowExecutionCanceledEventAttributes', childWorkflowExecutionCanceledEventAttributes_workflowType - The type of the child workflow execution.

$sel:initiatedEventId:ChildWorkflowExecutionCanceledEventAttributes', childWorkflowExecutionCanceledEventAttributes_initiatedEventId - The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution Decision to start this child workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

$sel:startedEventId:ChildWorkflowExecutionCanceledEventAttributes', childWorkflowExecutionCanceledEventAttributes_startedEventId - The ID of the ChildWorkflowExecutionStarted event recorded when this child workflow execution was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

childWorkflowExecutionCanceledEventAttributes_details :: Lens' ChildWorkflowExecutionCanceledEventAttributes (Maybe Text)

Details of the cancellation (if provided).

childWorkflowExecutionCanceledEventAttributes_workflowExecution :: Lens' ChildWorkflowExecutionCanceledEventAttributes WorkflowExecution

The child workflow execution that was canceled.

childWorkflowExecutionCanceledEventAttributes_workflowType :: Lens' ChildWorkflowExecutionCanceledEventAttributes WorkflowType

The type of the child workflow execution.

childWorkflowExecutionCanceledEventAttributes_initiatedEventId :: Lens' ChildWorkflowExecutionCanceledEventAttributes Integer

The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution Decision to start this child workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

childWorkflowExecutionCanceledEventAttributes_startedEventId :: Lens' ChildWorkflowExecutionCanceledEventAttributes Integer

The ID of the ChildWorkflowExecutionStarted event recorded when this child workflow execution was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.