Amazonka.SWF.Types.ChildWorkflowExecutionTimedOutEventAttributes

Description

Documentation

data ChildWorkflowExecutionTimedOutEventAttributes

Provides the details of the ChildWorkflowExecutionTimedOut event.

See: newChildWorkflowExecutionTimedOutEventAttributes smart constructor.

Constructors

ChildWorkflowExecutionTimedOutEventAttributes'

Fields

workflowExecution :: WorkflowExecution

The child workflow execution that timed out.

workflowType :: WorkflowType

The type of the child workflow execution.

timeoutType :: WorkflowExecutionTimeoutType

The type of the timeout that caused the child workflow execution to time out.

initiatedEventId :: Integer

The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution Decision to start this child workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

startedEventId :: Integer

The ID of the ChildWorkflowExecutionStarted event recorded when this child workflow execution was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newChildWorkflowExecutionTimedOutEventAttributes

Arguments

:: WorkflowExecution
$sel:workflowExecution:ChildWorkflowExecutionTimedOutEventAttributes'

-> WorkflowType
$sel:workflowType:ChildWorkflowExecutionTimedOutEventAttributes'

-> WorkflowExecutionTimeoutType
$sel:timeoutType:ChildWorkflowExecutionTimedOutEventAttributes'

-> Integer
$sel:initiatedEventId:ChildWorkflowExecutionTimedOutEventAttributes'

-> Integer
$sel:startedEventId:ChildWorkflowExecutionTimedOutEventAttributes'

-> ChildWorkflowExecutionTimedOutEventAttributes

Create a value of ChildWorkflowExecutionTimedOutEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workflowExecution:ChildWorkflowExecutionTimedOutEventAttributes', childWorkflowExecutionTimedOutEventAttributes_workflowExecution - The child workflow execution that timed out.

$sel:workflowType:ChildWorkflowExecutionTimedOutEventAttributes', childWorkflowExecutionTimedOutEventAttributes_workflowType - The type of the child workflow execution.

$sel:timeoutType:ChildWorkflowExecutionTimedOutEventAttributes', childWorkflowExecutionTimedOutEventAttributes_timeoutType - The type of the timeout that caused the child workflow execution to time out.

$sel:initiatedEventId:ChildWorkflowExecutionTimedOutEventAttributes', childWorkflowExecutionTimedOutEventAttributes_initiatedEventId - The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution Decision to start this child workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

$sel:startedEventId:ChildWorkflowExecutionTimedOutEventAttributes', childWorkflowExecutionTimedOutEventAttributes_startedEventId - The ID of the ChildWorkflowExecutionStarted event recorded when this child workflow execution was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

childWorkflowExecutionTimedOutEventAttributes_workflowExecution :: Lens' ChildWorkflowExecutionTimedOutEventAttributes WorkflowExecution

The child workflow execution that timed out.

childWorkflowExecutionTimedOutEventAttributes_workflowType :: Lens' ChildWorkflowExecutionTimedOutEventAttributes WorkflowType

The type of the child workflow execution.

childWorkflowExecutionTimedOutEventAttributes_timeoutType :: Lens' ChildWorkflowExecutionTimedOutEventAttributes WorkflowExecutionTimeoutType

The type of the timeout that caused the child workflow execution to time out.

childWorkflowExecutionTimedOutEventAttributes_initiatedEventId :: Lens' ChildWorkflowExecutionTimedOutEventAttributes Integer

The ID of the StartChildWorkflowExecutionInitiated event corresponding to the StartChildWorkflowExecution Decision to start this child workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

childWorkflowExecutionTimedOutEventAttributes_startedEventId :: Lens' ChildWorkflowExecutionTimedOutEventAttributes Integer

The ID of the ChildWorkflowExecutionStarted event recorded when this child workflow execution was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.