Amazonka.SWF.Types.ContinueAsNewWorkflowExecutionFailedEventAttributes

Description

Documentation

data ContinueAsNewWorkflowExecutionFailedEventAttributes

Provides the details of the ContinueAsNewWorkflowExecutionFailed event.

See: newContinueAsNewWorkflowExecutionFailedEventAttributes smart constructor.

Constructors

ContinueAsNewWorkflowExecutionFailedEventAttributes'

Fields

cause :: ContinueAsNewWorkflowExecutionFailedCause

The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

decisionTaskCompletedEventId :: Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the ContinueAsNewWorkflowExecution decision that started this execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newContinueAsNewWorkflowExecutionFailedEventAttributes

Arguments

:: ContinueAsNewWorkflowExecutionFailedCause
$sel:cause:ContinueAsNewWorkflowExecutionFailedEventAttributes'

-> Integer
$sel:decisionTaskCompletedEventId:ContinueAsNewWorkflowExecutionFailedEventAttributes'

-> ContinueAsNewWorkflowExecutionFailedEventAttributes

Create a value of ContinueAsNewWorkflowExecutionFailedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cause:ContinueAsNewWorkflowExecutionFailedEventAttributes', continueAsNewWorkflowExecutionFailedEventAttributes_cause - The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

$sel:decisionTaskCompletedEventId:ContinueAsNewWorkflowExecutionFailedEventAttributes', continueAsNewWorkflowExecutionFailedEventAttributes_decisionTaskCompletedEventId - The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the ContinueAsNewWorkflowExecution decision that started this execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

continueAsNewWorkflowExecutionFailedEventAttributes_cause :: Lens' ContinueAsNewWorkflowExecutionFailedEventAttributes ContinueAsNewWorkflowExecutionFailedCause

The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

continueAsNewWorkflowExecutionFailedEventAttributes_decisionTaskCompletedEventId :: Lens' ContinueAsNewWorkflowExecutionFailedEventAttributes Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the ContinueAsNewWorkflowExecution decision that started this execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.