Amazonka.SWF.Types.DecisionTaskScheduledEventAttributes

Description

Documentation

data DecisionTaskScheduledEventAttributes

Provides details about the DecisionTaskScheduled event.

See: newDecisionTaskScheduledEventAttributes smart constructor.

Constructors

DecisionTaskScheduledEventAttributes'

Fields

taskPriority :: Maybe Text

A task priority that, if set, specifies the priority for this decision task. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

startToCloseTimeout :: Maybe Text

The maximum duration for this decision task. The task is considered timed out if it doesn't completed within this duration.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

taskList :: TaskList

The name of the task list in which the decision task was scheduled.

newDecisionTaskScheduledEventAttributes

Arguments

:: TaskList
$sel:taskList:DecisionTaskScheduledEventAttributes'

-> DecisionTaskScheduledEventAttributes

Create a value of DecisionTaskScheduledEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskPriority:DecisionTaskScheduledEventAttributes', decisionTaskScheduledEventAttributes_taskPriority - A task priority that, if set, specifies the priority for this decision task. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

$sel:startToCloseTimeout:DecisionTaskScheduledEventAttributes', decisionTaskScheduledEventAttributes_startToCloseTimeout - The maximum duration for this decision task. The task is considered timed out if it doesn't completed within this duration.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

$sel:taskList:DecisionTaskScheduledEventAttributes', decisionTaskScheduledEventAttributes_taskList - The name of the task list in which the decision task was scheduled.

decisionTaskScheduledEventAttributes_taskPriority :: Lens' DecisionTaskScheduledEventAttributes (Maybe Text)

A task priority that, if set, specifies the priority for this decision task. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

decisionTaskScheduledEventAttributes_startToCloseTimeout :: Lens' DecisionTaskScheduledEventAttributes (Maybe Text)

The maximum duration for this decision task. The task is considered timed out if it doesn't completed within this duration.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

decisionTaskScheduledEventAttributes_taskList :: Lens' DecisionTaskScheduledEventAttributes TaskList

The name of the task list in which the decision task was scheduled.