Amazonka.SWF.Types.ExecutionTimeFilter

Description

Documentation

data ExecutionTimeFilter

Used to filter the workflow executions in visibility APIs by various time-based rules. Each parameter, if specified, defines a rule that must be satisfied by each returned query result. The parameter values are in the Unix Time format. For example: "oldestDate": 1325376070.

See: newExecutionTimeFilter smart constructor.

Constructors

ExecutionTimeFilter'

Fields

latestDate :: Maybe POSIX

Specifies the latest start or close date and time to return.

oldestDate :: POSIX

Specifies the oldest start or close date and time to return.

newExecutionTimeFilter

Arguments

:: UTCTime
$sel:oldestDate:ExecutionTimeFilter'

-> ExecutionTimeFilter

Create a value of ExecutionTimeFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestDate:ExecutionTimeFilter', executionTimeFilter_latestDate - Specifies the latest start or close date and time to return.

$sel:oldestDate:ExecutionTimeFilter', executionTimeFilter_oldestDate - Specifies the oldest start or close date and time to return.

executionTimeFilter_latestDate :: Lens' ExecutionTimeFilter (Maybe UTCTime)

Specifies the latest start or close date and time to return.

executionTimeFilter_oldestDate :: Lens' ExecutionTimeFilter UTCTime

Specifies the oldest start or close date and time to return.