Amazonka.SWF.Types.LambdaFunctionCompletedEventAttributes

Description

Documentation

data LambdaFunctionCompletedEventAttributes

Provides the details of the LambdaFunctionCompleted event. It isn't set for other event types.

See: newLambdaFunctionCompletedEventAttributes smart constructor.

Constructors

LambdaFunctionCompletedEventAttributes'

Fields

result :: Maybe Text

The results of the Lambda task.

scheduledEventId :: Integer

The ID of the LambdaFunctionScheduled event that was recorded when this Lambda task was scheduled. To help diagnose issues, use this information to trace back the chain of events leading up to this event.

startedEventId :: Integer

The ID of the LambdaFunctionStarted event recorded when this activity task started. To help diagnose issues, use this information to trace back the chain of events leading up to this event.

newLambdaFunctionCompletedEventAttributes

Arguments

:: Integer
$sel:scheduledEventId:LambdaFunctionCompletedEventAttributes'

-> Integer
$sel:startedEventId:LambdaFunctionCompletedEventAttributes'

-> LambdaFunctionCompletedEventAttributes

Create a value of LambdaFunctionCompletedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:result:LambdaFunctionCompletedEventAttributes', lambdaFunctionCompletedEventAttributes_result - The results of the Lambda task.

$sel:scheduledEventId:LambdaFunctionCompletedEventAttributes', lambdaFunctionCompletedEventAttributes_scheduledEventId - The ID of the LambdaFunctionScheduled event that was recorded when this Lambda task was scheduled. To help diagnose issues, use this information to trace back the chain of events leading up to this event.

$sel:startedEventId:LambdaFunctionCompletedEventAttributes', lambdaFunctionCompletedEventAttributes_startedEventId - The ID of the LambdaFunctionStarted event recorded when this activity task started. To help diagnose issues, use this information to trace back the chain of events leading up to this event.

lambdaFunctionCompletedEventAttributes_result :: Lens' LambdaFunctionCompletedEventAttributes (Maybe Text)

The results of the Lambda task.

lambdaFunctionCompletedEventAttributes_scheduledEventId :: Lens' LambdaFunctionCompletedEventAttributes Integer

The ID of the LambdaFunctionScheduled event that was recorded when this Lambda task was scheduled. To help diagnose issues, use this information to trace back the chain of events leading up to this event.

lambdaFunctionCompletedEventAttributes_startedEventId :: Lens' LambdaFunctionCompletedEventAttributes Integer

The ID of the LambdaFunctionStarted event recorded when this activity task started. To help diagnose issues, use this information to trace back the chain of events leading up to this event.