Amazonka.SWF.Types.RequestCancelExternalWorkflowExecutionFailedEventAttributes

Description

Documentation

data RequestCancelExternalWorkflowExecutionFailedEventAttributes

Provides the details of the RequestCancelExternalWorkflowExecutionFailed event.

See: newRequestCancelExternalWorkflowExecutionFailedEventAttributes smart constructor.

Constructors

RequestCancelExternalWorkflowExecutionFailedEventAttributes'

Fields

control :: Maybe Text

The data attached to the event that the decider can use in subsequent workflow tasks. This data isn't sent to the workflow execution.

runId :: Maybe Text

The runId of the external workflow execution.

workflowId :: Text

The workflowId of the external workflow to which the cancel request was to be delivered.

cause :: RequestCancelExternalWorkflowExecutionFailedCause

The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

initiatedEventId :: Integer

The ID of the RequestCancelExternalWorkflowExecutionInitiated event corresponding to the RequestCancelExternalWorkflowExecution decision to cancel this external workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

decisionTaskCompletedEventId :: Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the RequestCancelExternalWorkflowExecution decision for this cancellation request. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newRequestCancelExternalWorkflowExecutionFailedEventAttributes

Arguments

:: Text
$sel:workflowId:RequestCancelExternalWorkflowExecutionFailedEventAttributes'

-> RequestCancelExternalWorkflowExecutionFailedCause
$sel:cause:RequestCancelExternalWorkflowExecutionFailedEventAttributes'

-> Integer
$sel:initiatedEventId:RequestCancelExternalWorkflowExecutionFailedEventAttributes'

-> Integer
$sel:decisionTaskCompletedEventId:RequestCancelExternalWorkflowExecutionFailedEventAttributes'

-> RequestCancelExternalWorkflowExecutionFailedEventAttributes

Create a value of RequestCancelExternalWorkflowExecutionFailedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:control:RequestCancelExternalWorkflowExecutionFailedEventAttributes', requestCancelExternalWorkflowExecutionFailedEventAttributes_control - The data attached to the event that the decider can use in subsequent workflow tasks. This data isn't sent to the workflow execution.

$sel:runId:RequestCancelExternalWorkflowExecutionFailedEventAttributes', requestCancelExternalWorkflowExecutionFailedEventAttributes_runId - The runId of the external workflow execution.

$sel:workflowId:RequestCancelExternalWorkflowExecutionFailedEventAttributes', requestCancelExternalWorkflowExecutionFailedEventAttributes_workflowId - The workflowId of the external workflow to which the cancel request was to be delivered.

$sel:cause:RequestCancelExternalWorkflowExecutionFailedEventAttributes', requestCancelExternalWorkflowExecutionFailedEventAttributes_cause - The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

$sel:initiatedEventId:RequestCancelExternalWorkflowExecutionFailedEventAttributes', requestCancelExternalWorkflowExecutionFailedEventAttributes_initiatedEventId - The ID of the RequestCancelExternalWorkflowExecutionInitiated event corresponding to the RequestCancelExternalWorkflowExecution decision to cancel this external workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

$sel:decisionTaskCompletedEventId:RequestCancelExternalWorkflowExecutionFailedEventAttributes', requestCancelExternalWorkflowExecutionFailedEventAttributes_decisionTaskCompletedEventId - The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the RequestCancelExternalWorkflowExecution decision for this cancellation request. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

requestCancelExternalWorkflowExecutionFailedEventAttributes_control :: Lens' RequestCancelExternalWorkflowExecutionFailedEventAttributes (Maybe Text)

The data attached to the event that the decider can use in subsequent workflow tasks. This data isn't sent to the workflow execution.

requestCancelExternalWorkflowExecutionFailedEventAttributes_runId :: Lens' RequestCancelExternalWorkflowExecutionFailedEventAttributes (Maybe Text)

The runId of the external workflow execution.

requestCancelExternalWorkflowExecutionFailedEventAttributes_workflowId :: Lens' RequestCancelExternalWorkflowExecutionFailedEventAttributes Text

The workflowId of the external workflow to which the cancel request was to be delivered.

requestCancelExternalWorkflowExecutionFailedEventAttributes_cause :: Lens' RequestCancelExternalWorkflowExecutionFailedEventAttributes RequestCancelExternalWorkflowExecutionFailedCause

The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

requestCancelExternalWorkflowExecutionFailedEventAttributes_initiatedEventId :: Lens' RequestCancelExternalWorkflowExecutionFailedEventAttributes Integer

The ID of the RequestCancelExternalWorkflowExecutionInitiated event corresponding to the RequestCancelExternalWorkflowExecution decision to cancel this external workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

requestCancelExternalWorkflowExecutionFailedEventAttributes_decisionTaskCompletedEventId :: Lens' RequestCancelExternalWorkflowExecutionFailedEventAttributes Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the RequestCancelExternalWorkflowExecution decision for this cancellation request. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.