Amazonka.SWF.Types.ScheduleActivityTaskFailedEventAttributes

Description

Documentation

data ScheduleActivityTaskFailedEventAttributes

Provides the details of the ScheduleActivityTaskFailed event.

See: newScheduleActivityTaskFailedEventAttributes smart constructor.

Constructors

ScheduleActivityTaskFailedEventAttributes'

Fields

activityType :: ActivityType

The activity type provided in the ScheduleActivityTask decision that failed.

activityId :: Text

The activityId provided in the ScheduleActivityTask decision that failed.

cause :: ScheduleActivityTaskFailedCause

The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

decisionTaskCompletedEventId :: Integer

The ID of the DecisionTaskCompleted event corresponding to the decision that resulted in the scheduling of this activity task. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newScheduleActivityTaskFailedEventAttributes

Arguments

:: ActivityType
$sel:activityType:ScheduleActivityTaskFailedEventAttributes'

-> Text
$sel:activityId:ScheduleActivityTaskFailedEventAttributes'

-> ScheduleActivityTaskFailedCause
$sel:cause:ScheduleActivityTaskFailedEventAttributes'

-> Integer
$sel:decisionTaskCompletedEventId:ScheduleActivityTaskFailedEventAttributes'

-> ScheduleActivityTaskFailedEventAttributes

Create a value of ScheduleActivityTaskFailedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activityType:ScheduleActivityTaskFailedEventAttributes', scheduleActivityTaskFailedEventAttributes_activityType - The activity type provided in the ScheduleActivityTask decision that failed.

$sel:activityId:ScheduleActivityTaskFailedEventAttributes', scheduleActivityTaskFailedEventAttributes_activityId - The activityId provided in the ScheduleActivityTask decision that failed.

$sel:cause:ScheduleActivityTaskFailedEventAttributes', scheduleActivityTaskFailedEventAttributes_cause - The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

$sel:decisionTaskCompletedEventId:ScheduleActivityTaskFailedEventAttributes', scheduleActivityTaskFailedEventAttributes_decisionTaskCompletedEventId - The ID of the DecisionTaskCompleted event corresponding to the decision that resulted in the scheduling of this activity task. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

scheduleActivityTaskFailedEventAttributes_activityType :: Lens' ScheduleActivityTaskFailedEventAttributes ActivityType

The activity type provided in the ScheduleActivityTask decision that failed.

scheduleActivityTaskFailedEventAttributes_activityId :: Lens' ScheduleActivityTaskFailedEventAttributes Text

The activityId provided in the ScheduleActivityTask decision that failed.

scheduleActivityTaskFailedEventAttributes_cause :: Lens' ScheduleActivityTaskFailedEventAttributes ScheduleActivityTaskFailedCause

The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

If cause is set to OPERATION_NOT_PERMITTED, the decision failed because it lacked sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

scheduleActivityTaskFailedEventAttributes_decisionTaskCompletedEventId :: Lens' ScheduleActivityTaskFailedEventAttributes Integer

The ID of the DecisionTaskCompleted event corresponding to the decision that resulted in the scheduling of this activity task. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.