Amazonka.SWF.Types.StartChildWorkflowExecutionFailedEventAttributes

Description

Documentation

data StartChildWorkflowExecutionFailedEventAttributes

Provides the details of the StartChildWorkflowExecutionFailed event.

See: newStartChildWorkflowExecutionFailedEventAttributes smart constructor.

Constructors

StartChildWorkflowExecutionFailedEventAttributes'

Fields

control :: Maybe Text

The data attached to the event that the decider can use in subsequent workflow tasks. This data isn't sent to the child workflow execution.

workflowType :: WorkflowType

The workflow type provided in the StartChildWorkflowExecution Decision that failed.

cause :: StartChildWorkflowExecutionFailedCause

The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

When cause is set to OPERATION_NOT_PERMITTED, the decision fails because it lacks sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

workflowId :: Text

The workflowId of the child workflow execution.

initiatedEventId :: Integer

When the cause is WORKFLOW_ALREADY_RUNNING, initiatedEventId is the ID of the StartChildWorkflowExecutionInitiated event that corresponds to the StartChildWorkflowExecution Decision to start the workflow execution. You can use this information to diagnose problems by tracing back the chain of events leading up to this event.

When the cause isn't WORKFLOW_ALREADY_RUNNING, initiatedEventId is set to 0 because the StartChildWorkflowExecutionInitiated event doesn't exist.

decisionTaskCompletedEventId :: Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the StartChildWorkflowExecution Decision to request this child workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events.

newStartChildWorkflowExecutionFailedEventAttributes

Arguments

:: WorkflowType
$sel:workflowType:StartChildWorkflowExecutionFailedEventAttributes'

-> StartChildWorkflowExecutionFailedCause
$sel:cause:StartChildWorkflowExecutionFailedEventAttributes'

-> Text
$sel:workflowId:StartChildWorkflowExecutionFailedEventAttributes'

-> Integer
$sel:initiatedEventId:StartChildWorkflowExecutionFailedEventAttributes'

-> Integer
$sel:decisionTaskCompletedEventId:StartChildWorkflowExecutionFailedEventAttributes'

-> StartChildWorkflowExecutionFailedEventAttributes

Create a value of StartChildWorkflowExecutionFailedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:control:StartChildWorkflowExecutionFailedEventAttributes', startChildWorkflowExecutionFailedEventAttributes_control - The data attached to the event that the decider can use in subsequent workflow tasks. This data isn't sent to the child workflow execution.

$sel:workflowType:StartChildWorkflowExecutionFailedEventAttributes', startChildWorkflowExecutionFailedEventAttributes_workflowType - The workflow type provided in the StartChildWorkflowExecution Decision that failed.

$sel:cause:StartChildWorkflowExecutionFailedEventAttributes', startChildWorkflowExecutionFailedEventAttributes_cause - The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

When cause is set to OPERATION_NOT_PERMITTED, the decision fails because it lacks sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

$sel:workflowId:StartChildWorkflowExecutionFailedEventAttributes', startChildWorkflowExecutionFailedEventAttributes_workflowId - The workflowId of the child workflow execution.

$sel:initiatedEventId:StartChildWorkflowExecutionFailedEventAttributes', startChildWorkflowExecutionFailedEventAttributes_initiatedEventId - When the cause is WORKFLOW_ALREADY_RUNNING, initiatedEventId is the ID of the StartChildWorkflowExecutionInitiated event that corresponds to the StartChildWorkflowExecution Decision to start the workflow execution. You can use this information to diagnose problems by tracing back the chain of events leading up to this event.

When the cause isn't WORKFLOW_ALREADY_RUNNING, initiatedEventId is set to 0 because the StartChildWorkflowExecutionInitiated event doesn't exist.

$sel:decisionTaskCompletedEventId:StartChildWorkflowExecutionFailedEventAttributes', startChildWorkflowExecutionFailedEventAttributes_decisionTaskCompletedEventId - The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the StartChildWorkflowExecution Decision to request this child workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events.

startChildWorkflowExecutionFailedEventAttributes_control :: Lens' StartChildWorkflowExecutionFailedEventAttributes (Maybe Text)

The data attached to the event that the decider can use in subsequent workflow tasks. This data isn't sent to the child workflow execution.

startChildWorkflowExecutionFailedEventAttributes_workflowType :: Lens' StartChildWorkflowExecutionFailedEventAttributes WorkflowType

The workflow type provided in the StartChildWorkflowExecution Decision that failed.

startChildWorkflowExecutionFailedEventAttributes_cause :: Lens' StartChildWorkflowExecutionFailedEventAttributes StartChildWorkflowExecutionFailedCause

The cause of the failure. This information is generated by the system and can be useful for diagnostic purposes.

When cause is set to OPERATION_NOT_PERMITTED, the decision fails because it lacks sufficient permissions. For details and example IAM policies, see Using IAM to Manage Access to Amazon SWF Workflows in the Amazon SWF Developer Guide.

startChildWorkflowExecutionFailedEventAttributes_workflowId :: Lens' StartChildWorkflowExecutionFailedEventAttributes Text

The workflowId of the child workflow execution.

startChildWorkflowExecutionFailedEventAttributes_initiatedEventId :: Lens' StartChildWorkflowExecutionFailedEventAttributes Integer

When the cause is WORKFLOW_ALREADY_RUNNING, initiatedEventId is the ID of the StartChildWorkflowExecutionInitiated event that corresponds to the StartChildWorkflowExecution Decision to start the workflow execution. You can use this information to diagnose problems by tracing back the chain of events leading up to this event.

When the cause isn't WORKFLOW_ALREADY_RUNNING, initiatedEventId is set to 0 because the StartChildWorkflowExecutionInitiated event doesn't exist.

startChildWorkflowExecutionFailedEventAttributes_decisionTaskCompletedEventId :: Lens' StartChildWorkflowExecutionFailedEventAttributes Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the StartChildWorkflowExecution Decision to request this child workflow execution. This information can be useful for diagnosing problems by tracing back the chain of events.