Amazonka.SWF.Types.TimerCanceledEventAttributes

Description

Documentation

data TimerCanceledEventAttributes

Provides the details of the TimerCanceled event.

See: newTimerCanceledEventAttributes smart constructor.

Constructors

TimerCanceledEventAttributes'

Fields

timerId :: Text

The unique ID of the timer that was canceled.

startedEventId :: Integer

The ID of the TimerStarted event that was recorded when this timer was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

decisionTaskCompletedEventId :: Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the CancelTimer decision to cancel this timer. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newTimerCanceledEventAttributes

Arguments

:: Text
$sel:timerId:TimerCanceledEventAttributes'

-> Integer
$sel:startedEventId:TimerCanceledEventAttributes'

-> Integer
$sel:decisionTaskCompletedEventId:TimerCanceledEventAttributes'

-> TimerCanceledEventAttributes

Create a value of TimerCanceledEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timerId:TimerCanceledEventAttributes', timerCanceledEventAttributes_timerId - The unique ID of the timer that was canceled.

$sel:startedEventId:TimerCanceledEventAttributes', timerCanceledEventAttributes_startedEventId - The ID of the TimerStarted event that was recorded when this timer was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

$sel:decisionTaskCompletedEventId:TimerCanceledEventAttributes', timerCanceledEventAttributes_decisionTaskCompletedEventId - The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the CancelTimer decision to cancel this timer. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

timerCanceledEventAttributes_timerId :: Lens' TimerCanceledEventAttributes Text

The unique ID of the timer that was canceled.

timerCanceledEventAttributes_startedEventId :: Lens' TimerCanceledEventAttributes Integer

The ID of the TimerStarted event that was recorded when this timer was started. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

timerCanceledEventAttributes_decisionTaskCompletedEventId :: Lens' TimerCanceledEventAttributes Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the CancelTimer decision to cancel this timer. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.