Amazonka.SWF.Types.WorkflowExecutionCompletedEventAttributes

Description

Documentation

data WorkflowExecutionCompletedEventAttributes

Provides the details of the WorkflowExecutionCompleted event.

See: newWorkflowExecutionCompletedEventAttributes smart constructor.

Constructors

WorkflowExecutionCompletedEventAttributes'

Fields

result :: Maybe Text

The result produced by the workflow execution upon successful completion.

decisionTaskCompletedEventId :: Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the CompleteWorkflowExecution decision to complete this execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

newWorkflowExecutionCompletedEventAttributes

Arguments

:: Integer
$sel:decisionTaskCompletedEventId:WorkflowExecutionCompletedEventAttributes'

-> WorkflowExecutionCompletedEventAttributes

Create a value of WorkflowExecutionCompletedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:result:WorkflowExecutionCompletedEventAttributes', workflowExecutionCompletedEventAttributes_result - The result produced by the workflow execution upon successful completion.

$sel:decisionTaskCompletedEventId:WorkflowExecutionCompletedEventAttributes', workflowExecutionCompletedEventAttributes_decisionTaskCompletedEventId - The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the CompleteWorkflowExecution decision to complete this execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.

workflowExecutionCompletedEventAttributes_result :: Lens' WorkflowExecutionCompletedEventAttributes (Maybe Text)

The result produced by the workflow execution upon successful completion.

workflowExecutionCompletedEventAttributes_decisionTaskCompletedEventId :: Lens' WorkflowExecutionCompletedEventAttributes Integer

The ID of the DecisionTaskCompleted event corresponding to the decision task that resulted in the CompleteWorkflowExecution decision to complete this execution. This information can be useful for diagnosing problems by tracing back the chain of events leading up to this event.