Amazonka.SWF.Types.WorkflowExecutionConfiguration

Description

Documentation

data WorkflowExecutionConfiguration

The configuration settings for a workflow execution including timeout values, tasklist etc. These configuration settings are determined from the defaults specified when registering the workflow type and those specified when starting the workflow execution.

See: newWorkflowExecutionConfiguration smart constructor.

Constructors

WorkflowExecutionConfiguration'

Fields

lambdaRole :: Maybe Text

The IAM role attached to the child workflow execution.

taskPriority :: Maybe Text

The priority assigned to decision tasks for this workflow execution. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

taskStartToCloseTimeout :: Text

The maximum duration allowed for decision tasks for this workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

executionStartToCloseTimeout :: Text

The total duration for this workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

taskList :: TaskList

The task list used for the decision tasks generated for this workflow execution.

childPolicy :: ChildPolicy

The policy to use for the child workflow executions if this workflow execution is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

newWorkflowExecutionConfiguration

Arguments

:: Text
$sel:taskStartToCloseTimeout:WorkflowExecutionConfiguration'

-> Text
$sel:executionStartToCloseTimeout:WorkflowExecutionConfiguration'

-> TaskList
$sel:taskList:WorkflowExecutionConfiguration'

-> ChildPolicy
$sel:childPolicy:WorkflowExecutionConfiguration'

-> WorkflowExecutionConfiguration

Create a value of WorkflowExecutionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lambdaRole:WorkflowExecutionConfiguration', workflowExecutionConfiguration_lambdaRole - The IAM role attached to the child workflow execution.

$sel:taskPriority:WorkflowExecutionConfiguration', workflowExecutionConfiguration_taskPriority - The priority assigned to decision tasks for this workflow execution. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

$sel:taskStartToCloseTimeout:WorkflowExecutionConfiguration', workflowExecutionConfiguration_taskStartToCloseTimeout - The maximum duration allowed for decision tasks for this workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

$sel:executionStartToCloseTimeout:WorkflowExecutionConfiguration', workflowExecutionConfiguration_executionStartToCloseTimeout - The total duration for this workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

$sel:taskList:WorkflowExecutionConfiguration', workflowExecutionConfiguration_taskList - The task list used for the decision tasks generated for this workflow execution.

$sel:childPolicy:WorkflowExecutionConfiguration', workflowExecutionConfiguration_childPolicy - The policy to use for the child workflow executions if this workflow execution is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

workflowExecutionConfiguration_lambdaRole :: Lens' WorkflowExecutionConfiguration (Maybe Text)

The IAM role attached to the child workflow execution.

workflowExecutionConfiguration_taskPriority :: Lens' WorkflowExecutionConfiguration (Maybe Text)

The priority assigned to decision tasks for this workflow execution. Valid values are integers that range from Java's Integer.MIN_VALUE (-2147483648) to Integer.MAX_VALUE (2147483647). Higher numbers indicate higher priority.

For more information about setting task priority, see Setting Task Priority in the Amazon SWF Developer Guide.

workflowExecutionConfiguration_taskStartToCloseTimeout :: Lens' WorkflowExecutionConfiguration Text

The maximum duration allowed for decision tasks for this workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

workflowExecutionConfiguration_executionStartToCloseTimeout :: Lens' WorkflowExecutionConfiguration Text

The total duration for this workflow execution.

The duration is specified in seconds, an integer greater than or equal to 0. You can use NONE to specify unlimited duration.

workflowExecutionConfiguration_taskList :: Lens' WorkflowExecutionConfiguration TaskList

The task list used for the decision tasks generated for this workflow execution.

workflowExecutionConfiguration_childPolicy :: Lens' WorkflowExecutionConfiguration ChildPolicy

The policy to use for the child workflow executions if this workflow execution is terminated, by calling the TerminateWorkflowExecution action explicitly or due to an expired timeout.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.