Amazonka.SWF.Types.WorkflowExecutionInfos

Description

Documentation

data WorkflowExecutionInfos

Contains a paginated list of information about workflow executions.

See: newWorkflowExecutionInfos smart constructor.

Constructors

WorkflowExecutionInfos'

Fields

nextPageToken :: Maybe Text

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

executionInfos :: [WorkflowExecutionInfo]

The list of workflow information structures.

newWorkflowExecutionInfos :: WorkflowExecutionInfos

Create a value of WorkflowExecutionInfos with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:WorkflowExecutionInfos', workflowExecutionInfos_nextPageToken - If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

$sel:executionInfos:WorkflowExecutionInfos', workflowExecutionInfos_executionInfos - The list of workflow information structures.

workflowExecutionInfos_nextPageToken :: Lens' WorkflowExecutionInfos (Maybe Text)

If a NextPageToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextPageToken. Keep all other arguments unchanged.

The configured maximumPageSize determines how many results can be returned in a single call.

workflowExecutionInfos_executionInfos :: Lens' WorkflowExecutionInfos [WorkflowExecutionInfo]

The list of workflow information structures.