Amazonka.SWF.Types.WorkflowExecutionTerminatedEventAttributes

Description

Documentation

data WorkflowExecutionTerminatedEventAttributes

Provides the details of the WorkflowExecutionTerminated event.

See: newWorkflowExecutionTerminatedEventAttributes smart constructor.

Constructors

WorkflowExecutionTerminatedEventAttributes'

Fields

cause :: Maybe WorkflowExecutionTerminatedCause

If set, indicates that the workflow execution was automatically terminated, and specifies the cause. This happens if the parent workflow execution times out or is terminated and the child policy is set to terminate child executions.

reason :: Maybe Text

The reason provided for the termination.

details :: Maybe Text

The details provided for the termination.

childPolicy :: ChildPolicy

The policy used for the child workflow executions of this workflow execution.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

newWorkflowExecutionTerminatedEventAttributes

Arguments

:: ChildPolicy
$sel:childPolicy:WorkflowExecutionTerminatedEventAttributes'

-> WorkflowExecutionTerminatedEventAttributes

Create a value of WorkflowExecutionTerminatedEventAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cause:WorkflowExecutionTerminatedEventAttributes', workflowExecutionTerminatedEventAttributes_cause - If set, indicates that the workflow execution was automatically terminated, and specifies the cause. This happens if the parent workflow execution times out or is terminated and the child policy is set to terminate child executions.

$sel:reason:WorkflowExecutionTerminatedEventAttributes', workflowExecutionTerminatedEventAttributes_reason - The reason provided for the termination.

$sel:details:WorkflowExecutionTerminatedEventAttributes', workflowExecutionTerminatedEventAttributes_details - The details provided for the termination.

$sel:childPolicy:WorkflowExecutionTerminatedEventAttributes', workflowExecutionTerminatedEventAttributes_childPolicy - The policy used for the child workflow executions of this workflow execution.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.

workflowExecutionTerminatedEventAttributes_cause :: Lens' WorkflowExecutionTerminatedEventAttributes (Maybe WorkflowExecutionTerminatedCause)

If set, indicates that the workflow execution was automatically terminated, and specifies the cause. This happens if the parent workflow execution times out or is terminated and the child policy is set to terminate child executions.

workflowExecutionTerminatedEventAttributes_reason :: Lens' WorkflowExecutionTerminatedEventAttributes (Maybe Text)

The reason provided for the termination.

workflowExecutionTerminatedEventAttributes_details :: Lens' WorkflowExecutionTerminatedEventAttributes (Maybe Text)

The details provided for the termination.

workflowExecutionTerminatedEventAttributes_childPolicy :: Lens' WorkflowExecutionTerminatedEventAttributes ChildPolicy

The policy used for the child workflow executions of this workflow execution.

The supported child policies are:

TERMINATE – The child executions are terminated.

REQUEST_CANCEL – A request to cancel is attempted for each child execution by recording a WorkflowExecutionCancelRequested event in its history. It is up to the decider to take appropriate actions when it receives an execution history with this event.

ABANDON – No action is taken. The child executions continue to run.