Amazonka.SageMaker.CreateArtifact

Description

Creates an artifact. An artifact is a lineage tracking entity that represents a URI addressable object or data. Some examples are the S3 URI of a dataset and the ECR registry path of an image. For more information, see Amazon SageMaker ML Lineage Tracking.

Creating a Request

data CreateArtifact

See: newCreateArtifact smart constructor.

Constructors

CreateArtifact'

Fields

metadataProperties :: Maybe MetadataProperties

artifactName :: Maybe Text

The name of the artifact. Must be unique to your account in an Amazon Web Services Region.

tags :: Maybe [Tag]

A list of tags to apply to the artifact.

properties :: Maybe (HashMap Text Text)

A list of properties to add to the artifact.

source :: ArtifactSource

The ID, ID type, and URI of the source.

artifactType :: Text

The artifact type.

newCreateArtifact

Arguments

:: ArtifactSource
$sel:source:CreateArtifact'

-> Text
$sel:artifactType:CreateArtifact'

-> CreateArtifact

Create a value of CreateArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadataProperties:CreateArtifact', createArtifact_metadataProperties - Undocumented member.

$sel:artifactName:CreateArtifact', createArtifact_artifactName - The name of the artifact. Must be unique to your account in an Amazon Web Services Region.

$sel:tags:CreateArtifact', createArtifact_tags - A list of tags to apply to the artifact.

$sel:properties:CreateArtifact', createArtifact_properties - A list of properties to add to the artifact.

$sel:source:CreateArtifact', createArtifact_source - The ID, ID type, and URI of the source.

$sel:artifactType:CreateArtifact', createArtifact_artifactType - The artifact type.

Request Lenses

createArtifact_metadataProperties :: Lens' CreateArtifact (Maybe MetadataProperties)

Undocumented member.

createArtifact_artifactName :: Lens' CreateArtifact (Maybe Text)

The name of the artifact. Must be unique to your account in an Amazon Web Services Region.

createArtifact_tags :: Lens' CreateArtifact (Maybe [Tag])

A list of tags to apply to the artifact.

createArtifact_properties :: Lens' CreateArtifact (Maybe (HashMap Text Text))

A list of properties to add to the artifact.

createArtifact_source :: Lens' CreateArtifact ArtifactSource

The ID, ID type, and URI of the source.

createArtifact_artifactType :: Lens' CreateArtifact Text

The artifact type.

Destructuring the Response

data CreateArtifactResponse

See: newCreateArtifactResponse smart constructor.

Constructors

CreateArtifactResponse'

Fields

artifactArn :: Maybe Text

The Amazon Resource Name (ARN) of the artifact.

httpStatus :: Int

The response's http status code.

newCreateArtifactResponse

Arguments

:: Int
$sel:httpStatus:CreateArtifactResponse'

-> CreateArtifactResponse

Create a value of CreateArtifactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:artifactArn:CreateArtifactResponse', createArtifactResponse_artifactArn - The Amazon Resource Name (ARN) of the artifact.

$sel:httpStatus:CreateArtifactResponse', createArtifactResponse_httpStatus - The response's http status code.

Response Lenses

createArtifactResponse_artifactArn :: Lens' CreateArtifactResponse (Maybe Text)

The Amazon Resource Name (ARN) of the artifact.

createArtifactResponse_httpStatus :: Lens' CreateArtifactResponse Int

The response's http status code.