Amazonka.SageMaker.CreateCompilationJob

Description

Starts a model compilation job. After the model has been compiled, Amazon SageMaker saves the resulting model artifacts to an Amazon Simple Storage Service (Amazon S3) bucket that you specify.

If you choose to host your model using Amazon SageMaker hosting services, you can use the resulting model artifacts as part of the model. You can also use the artifacts with Amazon Web Services IoT Greengrass. In that case, deploy them as an ML resource.

In the request body, you provide the following:

A name for the compilation job

Information about the input model artifacts

The output location for the compiled model and the device (target) that the model runs on

The Amazon Resource Name (ARN) of the IAM role that Amazon SageMaker assumes to perform the model compilation job.

You can also provide a Tag to track the model compilation job's resource use and costs. The response body contains the CompilationJobArn for the compiled job.

To stop a model compilation job, use StopCompilationJob. To get information about a particular model compilation job, use DescribeCompilationJob. To get information about multiple model compilation jobs, use ListCompilationJobs.

Creating a Request

data CreateCompilationJob

See: newCreateCompilationJob smart constructor.

Constructors

CreateCompilationJob'

Fields

vpcConfig :: Maybe NeoVpcConfig

A VpcConfig object that specifies the VPC that you want your compilation job to connect to. Control access to your models by configuring the VPC. For more information, see Protect Compilation Jobs by Using an Amazon Virtual Private Cloud.

tags :: Maybe [Tag]

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

compilationJobName :: Text

A name for the model compilation job. The name must be unique within the Amazon Web Services Region and within your Amazon Web Services account.

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that enables Amazon SageMaker to perform tasks on your behalf.

During model compilation, Amazon SageMaker needs your permission to:

Read input data from an S3 bucket

Write model artifacts to an S3 bucket

Write logs to Amazon CloudWatch Logs

Publish metrics to Amazon CloudWatch

You grant permissions for all of these tasks to an IAM role. To pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission. For more information, see Amazon SageMaker Roles.

inputConfig :: InputConfig

Provides information about the location of input model artifacts, the name and shape of the expected data inputs, and the framework in which the model was trained.

outputConfig :: OutputConfig

Provides information about the output location for the compiled model and the target device the model runs on.

stoppingCondition :: StoppingCondition

Specifies a limit to how long a model compilation job can run. When the job reaches the time limit, Amazon SageMaker ends the compilation job. Use this API to cap model training costs.

newCreateCompilationJob

Arguments

:: Text
$sel:compilationJobName:CreateCompilationJob'

-> Text
$sel:roleArn:CreateCompilationJob'

-> InputConfig
$sel:inputConfig:CreateCompilationJob'

-> OutputConfig
$sel:outputConfig:CreateCompilationJob'

-> StoppingCondition
$sel:stoppingCondition:CreateCompilationJob'

-> CreateCompilationJob

Create a value of CreateCompilationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:CreateCompilationJob', createCompilationJob_vpcConfig - A VpcConfig object that specifies the VPC that you want your compilation job to connect to. Control access to your models by configuring the VPC. For more information, see Protect Compilation Jobs by Using an Amazon Virtual Private Cloud.

$sel:tags:CreateCompilationJob', createCompilationJob_tags - An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

$sel:compilationJobName:CreateCompilationJob', createCompilationJob_compilationJobName - A name for the model compilation job. The name must be unique within the Amazon Web Services Region and within your Amazon Web Services account.

$sel:roleArn:CreateCompilationJob', createCompilationJob_roleArn - The Amazon Resource Name (ARN) of an IAM role that enables Amazon SageMaker to perform tasks on your behalf.

During model compilation, Amazon SageMaker needs your permission to:

Read input data from an S3 bucket

Write model artifacts to an S3 bucket

Write logs to Amazon CloudWatch Logs

Publish metrics to Amazon CloudWatch

You grant permissions for all of these tasks to an IAM role. To pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission. For more information, see Amazon SageMaker Roles.

$sel:inputConfig:CreateCompilationJob', createCompilationJob_inputConfig - Provides information about the location of input model artifacts, the name and shape of the expected data inputs, and the framework in which the model was trained.

$sel:outputConfig:CreateCompilationJob', createCompilationJob_outputConfig - Provides information about the output location for the compiled model and the target device the model runs on.

$sel:stoppingCondition:CreateCompilationJob', createCompilationJob_stoppingCondition - Specifies a limit to how long a model compilation job can run. When the job reaches the time limit, Amazon SageMaker ends the compilation job. Use this API to cap model training costs.

Request Lenses

createCompilationJob_vpcConfig :: Lens' CreateCompilationJob (Maybe NeoVpcConfig)

A VpcConfig object that specifies the VPC that you want your compilation job to connect to. Control access to your models by configuring the VPC. For more information, see Protect Compilation Jobs by Using an Amazon Virtual Private Cloud.

createCompilationJob_tags :: Lens' CreateCompilationJob (Maybe [Tag])

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

createCompilationJob_compilationJobName :: Lens' CreateCompilationJob Text

A name for the model compilation job. The name must be unique within the Amazon Web Services Region and within your Amazon Web Services account.

createCompilationJob_roleArn :: Lens' CreateCompilationJob Text

The Amazon Resource Name (ARN) of an IAM role that enables Amazon SageMaker to perform tasks on your behalf.

During model compilation, Amazon SageMaker needs your permission to:

Read input data from an S3 bucket

Write model artifacts to an S3 bucket

Write logs to Amazon CloudWatch Logs

Publish metrics to Amazon CloudWatch

You grant permissions for all of these tasks to an IAM role. To pass this role to Amazon SageMaker, the caller of this API must have the iam:PassRole permission. For more information, see Amazon SageMaker Roles.

createCompilationJob_inputConfig :: Lens' CreateCompilationJob InputConfig

Provides information about the location of input model artifacts, the name and shape of the expected data inputs, and the framework in which the model was trained.

createCompilationJob_outputConfig :: Lens' CreateCompilationJob OutputConfig

Provides information about the output location for the compiled model and the target device the model runs on.

createCompilationJob_stoppingCondition :: Lens' CreateCompilationJob StoppingCondition

Specifies a limit to how long a model compilation job can run. When the job reaches the time limit, Amazon SageMaker ends the compilation job. Use this API to cap model training costs.

Destructuring the Response

data CreateCompilationJobResponse

See: newCreateCompilationJobResponse smart constructor.

Constructors

CreateCompilationJobResponse'

Fields

httpStatus :: Int

The response's http status code.

compilationJobArn :: Text

If the action is successful, the service sends back an HTTP 200 response. Amazon SageMaker returns the following data in JSON format:

CompilationJobArn: The Amazon Resource Name (ARN) of the compiled job.

newCreateCompilationJobResponse

Arguments

:: Int
$sel:httpStatus:CreateCompilationJobResponse'

-> Text
$sel:compilationJobArn:CreateCompilationJobResponse'

-> CreateCompilationJobResponse

Create a value of CreateCompilationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateCompilationJobResponse', createCompilationJobResponse_httpStatus - The response's http status code.

$sel:compilationJobArn:CreateCompilationJobResponse', createCompilationJobResponse_compilationJobArn - If the action is successful, the service sends back an HTTP 200 response. Amazon SageMaker returns the following data in JSON format:

CompilationJobArn: The Amazon Resource Name (ARN) of the compiled job.

Response Lenses

createCompilationJobResponse_httpStatus :: Lens' CreateCompilationJobResponse Int

The response's http status code.

createCompilationJobResponse_compilationJobArn :: Lens' CreateCompilationJobResponse Text

If the action is successful, the service sends back an HTTP 200 response. Amazon SageMaker returns the following data in JSON format:

CompilationJobArn: The Amazon Resource Name (ARN) of the compiled job.