Amazonka.SageMaker.CreateHumanTaskUi

Description

Defines the settings you will use for the human review workflow user interface. Reviewers will see a three-panel interface with an instruction area, the item to review, and an input area.

Creating a Request

data CreateHumanTaskUi

See: newCreateHumanTaskUi smart constructor.

Constructors

CreateHumanTaskUi'

Fields

tags :: Maybe [Tag]

An array of key-value pairs that contain metadata to help you categorize and organize a human review workflow user interface. Each tag consists of a key and a value, both of which you define.

humanTaskUiName :: Text

The name of the user interface you are creating.

uiTemplate :: UiTemplate

newCreateHumanTaskUi

Arguments

:: Text
$sel:humanTaskUiName:CreateHumanTaskUi'

-> UiTemplate
$sel:uiTemplate:CreateHumanTaskUi'

-> CreateHumanTaskUi

Create a value of CreateHumanTaskUi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateHumanTaskUi', createHumanTaskUi_tags - An array of key-value pairs that contain metadata to help you categorize and organize a human review workflow user interface. Each tag consists of a key and a value, both of which you define.

$sel:humanTaskUiName:CreateHumanTaskUi', createHumanTaskUi_humanTaskUiName - The name of the user interface you are creating.

$sel:uiTemplate:CreateHumanTaskUi', createHumanTaskUi_uiTemplate - Undocumented member.

Request Lenses

createHumanTaskUi_tags :: Lens' CreateHumanTaskUi (Maybe [Tag])

An array of key-value pairs that contain metadata to help you categorize and organize a human review workflow user interface. Each tag consists of a key and a value, both of which you define.

createHumanTaskUi_humanTaskUiName :: Lens' CreateHumanTaskUi Text

The name of the user interface you are creating.

createHumanTaskUi_uiTemplate :: Lens' CreateHumanTaskUi UiTemplate

Undocumented member.

Destructuring the Response

data CreateHumanTaskUiResponse

See: newCreateHumanTaskUiResponse smart constructor.

Constructors

CreateHumanTaskUiResponse'

Fields

httpStatus :: Int

The response's http status code.

humanTaskUiArn :: Text

The Amazon Resource Name (ARN) of the human review workflow user interface you create.

newCreateHumanTaskUiResponse

Arguments

:: Int
$sel:httpStatus:CreateHumanTaskUiResponse'

-> Text
$sel:humanTaskUiArn:CreateHumanTaskUiResponse'

-> CreateHumanTaskUiResponse

Create a value of CreateHumanTaskUiResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateHumanTaskUiResponse', createHumanTaskUiResponse_httpStatus - The response's http status code.

$sel:humanTaskUiArn:CreateHumanTaskUiResponse', createHumanTaskUiResponse_humanTaskUiArn - The Amazon Resource Name (ARN) of the human review workflow user interface you create.

Response Lenses

createHumanTaskUiResponse_httpStatus :: Lens' CreateHumanTaskUiResponse Int

The response's http status code.

createHumanTaskUiResponse_humanTaskUiArn :: Lens' CreateHumanTaskUiResponse Text

The Amazon Resource Name (ARN) of the human review workflow user interface you create.