Amazonka.SageMaker.CreateImage

Description

Creates a custom SageMaker image. A SageMaker image is a set of image versions. Each image version represents a container image stored in Amazon Container Registry (ECR). For more information, see Bring your own SageMaker image.

Creating a Request

data CreateImage

See: newCreateImage smart constructor.

Constructors

CreateImage'

Fields

displayName :: Maybe Text

The display name of the image. If not provided, ImageName is displayed.

description :: Maybe Text

The description of the image.

tags :: Maybe [Tag]

A list of tags to apply to the image.

imageName :: Text

The name of the image. Must be unique to your account.

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that enables Amazon SageMaker to perform tasks on your behalf.

newCreateImage

Arguments

:: Text
$sel:imageName:CreateImage'

-> Text
$sel:roleArn:CreateImage'

-> CreateImage

Create a value of CreateImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:displayName:CreateImage', createImage_displayName - The display name of the image. If not provided, ImageName is displayed.

$sel:description:CreateImage', createImage_description - The description of the image.

$sel:tags:CreateImage', createImage_tags - A list of tags to apply to the image.

$sel:imageName:CreateImage', createImage_imageName - The name of the image. Must be unique to your account.

$sel:roleArn:CreateImage', createImage_roleArn - The Amazon Resource Name (ARN) of an IAM role that enables Amazon SageMaker to perform tasks on your behalf.

Request Lenses

createImage_displayName :: Lens' CreateImage (Maybe Text)

The display name of the image. If not provided, ImageName is displayed.

createImage_description :: Lens' CreateImage (Maybe Text)

The description of the image.

createImage_tags :: Lens' CreateImage (Maybe [Tag])

A list of tags to apply to the image.

createImage_imageName :: Lens' CreateImage Text

The name of the image. Must be unique to your account.

createImage_roleArn :: Lens' CreateImage Text

The Amazon Resource Name (ARN) of an IAM role that enables Amazon SageMaker to perform tasks on your behalf.

Destructuring the Response

data CreateImageResponse

See: newCreateImageResponse smart constructor.

Constructors

CreateImageResponse'

Fields

imageArn :: Maybe Text

The Amazon Resource Name (ARN) of the image.

httpStatus :: Int

The response's http status code.

newCreateImageResponse

Arguments

:: Int
$sel:httpStatus:CreateImageResponse'

-> CreateImageResponse

Create a value of CreateImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageArn:CreateImageResponse', createImageResponse_imageArn - The Amazon Resource Name (ARN) of the image.

$sel:httpStatus:CreateImageResponse', createImageResponse_httpStatus - The response's http status code.

Response Lenses

createImageResponse_imageArn :: Lens' CreateImageResponse (Maybe Text)

The Amazon Resource Name (ARN) of the image.

createImageResponse_httpStatus :: Lens' CreateImageResponse Int

The response's http status code.