Amazonka.SageMaker.CreateModelQualityJobDefinition

Description

Creates a definition for a job that monitors model quality and drift. For information about model monitor, see Amazon SageMaker Model Monitor.

Creating a Request

data CreateModelQualityJobDefinition

See: newCreateModelQualityJobDefinition smart constructor.

Constructors

CreateModelQualityJobDefinition'

Fields

modelQualityBaselineConfig :: Maybe ModelQualityBaselineConfig

Specifies the constraints and baselines for the monitoring job.

stoppingCondition :: Maybe MonitoringStoppingCondition

networkConfig :: Maybe MonitoringNetworkConfig

Specifies the network configuration for the monitoring job.

tags :: Maybe [Tag]

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

jobDefinitionName :: Text

The name of the monitoring job definition.

modelQualityAppSpecification :: ModelQualityAppSpecification

The container that runs the monitoring job.

modelQualityJobInput :: ModelQualityJobInput

A list of the inputs that are monitored. Currently endpoints are supported.

modelQualityJobOutputConfig :: MonitoringOutputConfig

jobResources :: MonitoringResources

roleArn :: Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

newCreateModelQualityJobDefinition

Arguments

:: Text
$sel:jobDefinitionName:CreateModelQualityJobDefinition'

-> ModelQualityAppSpecification
$sel:modelQualityAppSpecification:CreateModelQualityJobDefinition'

-> ModelQualityJobInput
$sel:modelQualityJobInput:CreateModelQualityJobDefinition'

-> MonitoringOutputConfig
$sel:modelQualityJobOutputConfig:CreateModelQualityJobDefinition'

-> MonitoringResources
$sel:jobResources:CreateModelQualityJobDefinition'

-> Text
$sel:roleArn:CreateModelQualityJobDefinition'

-> CreateModelQualityJobDefinition

Create a value of CreateModelQualityJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelQualityBaselineConfig:CreateModelQualityJobDefinition', createModelQualityJobDefinition_modelQualityBaselineConfig - Specifies the constraints and baselines for the monitoring job.

$sel:stoppingCondition:CreateModelQualityJobDefinition', createModelQualityJobDefinition_stoppingCondition - Undocumented member.

$sel:networkConfig:CreateModelQualityJobDefinition', createModelQualityJobDefinition_networkConfig - Specifies the network configuration for the monitoring job.

$sel:tags:CreateModelQualityJobDefinition', createModelQualityJobDefinition_tags - (Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

$sel:jobDefinitionName:CreateModelQualityJobDefinition', createModelQualityJobDefinition_jobDefinitionName - The name of the monitoring job definition.

$sel:modelQualityAppSpecification:CreateModelQualityJobDefinition', createModelQualityJobDefinition_modelQualityAppSpecification - The container that runs the monitoring job.

$sel:modelQualityJobInput:CreateModelQualityJobDefinition', createModelQualityJobDefinition_modelQualityJobInput - A list of the inputs that are monitored. Currently endpoints are supported.

$sel:modelQualityJobOutputConfig:CreateModelQualityJobDefinition', createModelQualityJobDefinition_modelQualityJobOutputConfig - Undocumented member.

$sel:jobResources:CreateModelQualityJobDefinition', createModelQualityJobDefinition_jobResources - Undocumented member.

$sel:roleArn:CreateModelQualityJobDefinition', createModelQualityJobDefinition_roleArn - The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

Request Lenses

createModelQualityJobDefinition_modelQualityBaselineConfig :: Lens' CreateModelQualityJobDefinition (Maybe ModelQualityBaselineConfig)

Specifies the constraints and baselines for the monitoring job.

createModelQualityJobDefinition_stoppingCondition :: Lens' CreateModelQualityJobDefinition (Maybe MonitoringStoppingCondition)

Undocumented member.

createModelQualityJobDefinition_networkConfig :: Lens' CreateModelQualityJobDefinition (Maybe MonitoringNetworkConfig)

Specifies the network configuration for the monitoring job.

createModelQualityJobDefinition_tags :: Lens' CreateModelQualityJobDefinition (Maybe [Tag])

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

createModelQualityJobDefinition_jobDefinitionName :: Lens' CreateModelQualityJobDefinition Text

The name of the monitoring job definition.

createModelQualityJobDefinition_modelQualityAppSpecification :: Lens' CreateModelQualityJobDefinition ModelQualityAppSpecification

The container that runs the monitoring job.

createModelQualityJobDefinition_modelQualityJobInput :: Lens' CreateModelQualityJobDefinition ModelQualityJobInput

A list of the inputs that are monitored. Currently endpoints are supported.

createModelQualityJobDefinition_modelQualityJobOutputConfig :: Lens' CreateModelQualityJobDefinition MonitoringOutputConfig

Undocumented member.

createModelQualityJobDefinition_jobResources :: Lens' CreateModelQualityJobDefinition MonitoringResources

Undocumented member.

createModelQualityJobDefinition_roleArn :: Lens' CreateModelQualityJobDefinition Text

The Amazon Resource Name (ARN) of an IAM role that Amazon SageMaker can assume to perform tasks on your behalf.

Destructuring the Response

data CreateModelQualityJobDefinitionResponse

See: newCreateModelQualityJobDefinitionResponse smart constructor.

Constructors

CreateModelQualityJobDefinitionResponse'

Fields

httpStatus :: Int

The response's http status code.

jobDefinitionArn :: Text

The Amazon Resource Name (ARN) of the model quality monitoring job.

newCreateModelQualityJobDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateModelQualityJobDefinitionResponse'

-> Text
$sel:jobDefinitionArn:CreateModelQualityJobDefinitionResponse'

-> CreateModelQualityJobDefinitionResponse

Create a value of CreateModelQualityJobDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateModelQualityJobDefinitionResponse', createModelQualityJobDefinitionResponse_httpStatus - The response's http status code.

$sel:jobDefinitionArn:CreateModelQualityJobDefinitionResponse', createModelQualityJobDefinitionResponse_jobDefinitionArn - The Amazon Resource Name (ARN) of the model quality monitoring job.

Response Lenses

createModelQualityJobDefinitionResponse_httpStatus :: Lens' CreateModelQualityJobDefinitionResponse Int

The response's http status code.

createModelQualityJobDefinitionResponse_jobDefinitionArn :: Lens' CreateModelQualityJobDefinitionResponse Text

The Amazon Resource Name (ARN) of the model quality monitoring job.