Amazonka.SageMaker.CreateMonitoringSchedule

Description

Creates a schedule that regularly starts Amazon SageMaker Processing Jobs to monitor the data captured for an Amazon SageMaker Endoint.

Creating a Request

data CreateMonitoringSchedule

See: newCreateMonitoringSchedule smart constructor.

Constructors

CreateMonitoringSchedule'

Fields

tags :: Maybe [Tag]

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

monitoringScheduleName :: Text

The name of the monitoring schedule. The name must be unique within an Amazon Web Services Region within an Amazon Web Services account.

monitoringScheduleConfig :: MonitoringScheduleConfig

The configuration object that specifies the monitoring schedule and defines the monitoring job.

newCreateMonitoringSchedule

Arguments

:: Text
$sel:monitoringScheduleName:CreateMonitoringSchedule'

-> MonitoringScheduleConfig
$sel:monitoringScheduleConfig:CreateMonitoringSchedule'

-> CreateMonitoringSchedule

Create a value of CreateMonitoringSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateMonitoringSchedule', createMonitoringSchedule_tags - (Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

$sel:monitoringScheduleName:CreateMonitoringSchedule', createMonitoringSchedule_monitoringScheduleName - The name of the monitoring schedule. The name must be unique within an Amazon Web Services Region within an Amazon Web Services account.

$sel:monitoringScheduleConfig:CreateMonitoringSchedule', createMonitoringSchedule_monitoringScheduleConfig - The configuration object that specifies the monitoring schedule and defines the monitoring job.

Request Lenses

createMonitoringSchedule_tags :: Lens' CreateMonitoringSchedule (Maybe [Tag])

(Optional) An array of key-value pairs. For more information, see Using Cost Allocation Tags in the Amazon Web Services Billing and Cost Management User Guide.

createMonitoringSchedule_monitoringScheduleName :: Lens' CreateMonitoringSchedule Text

The name of the monitoring schedule. The name must be unique within an Amazon Web Services Region within an Amazon Web Services account.

createMonitoringSchedule_monitoringScheduleConfig :: Lens' CreateMonitoringSchedule MonitoringScheduleConfig

The configuration object that specifies the monitoring schedule and defines the monitoring job.

Destructuring the Response

data CreateMonitoringScheduleResponse

See: newCreateMonitoringScheduleResponse smart constructor.

Constructors

CreateMonitoringScheduleResponse'

Fields

httpStatus :: Int

The response's http status code.

monitoringScheduleArn :: Text

The Amazon Resource Name (ARN) of the monitoring schedule.

newCreateMonitoringScheduleResponse

Arguments

:: Int
$sel:httpStatus:CreateMonitoringScheduleResponse'

-> Text
$sel:monitoringScheduleArn:CreateMonitoringScheduleResponse'

-> CreateMonitoringScheduleResponse

Create a value of CreateMonitoringScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateMonitoringScheduleResponse', createMonitoringScheduleResponse_httpStatus - The response's http status code.

$sel:monitoringScheduleArn:CreateMonitoringScheduleResponse', createMonitoringScheduleResponse_monitoringScheduleArn - The Amazon Resource Name (ARN) of the monitoring schedule.

Response Lenses

createMonitoringScheduleResponse_httpStatus :: Lens' CreateMonitoringScheduleResponse Int

The response's http status code.

createMonitoringScheduleResponse_monitoringScheduleArn :: Lens' CreateMonitoringScheduleResponse Text

The Amazon Resource Name (ARN) of the monitoring schedule.