Amazonka.SageMaker.CreateStudioLifecycleConfig

Description

Creates a new Studio Lifecycle Configuration.

Creating a Request

data CreateStudioLifecycleConfig

See: newCreateStudioLifecycleConfig smart constructor.

Constructors

CreateStudioLifecycleConfig'

Fields

tags :: Maybe [Tag]

Tags to be associated with the Lifecycle Configuration. Each tag consists of a key and an optional value. Tag keys must be unique per resource. Tags are searchable using the Search API.

studioLifecycleConfigName :: Text

The name of the Studio Lifecycle Configuration to create.

studioLifecycleConfigContent :: Text

The content of your Studio Lifecycle Configuration script. This content must be base64 encoded.

studioLifecycleConfigAppType :: StudioLifecycleConfigAppType

The App type that the Lifecycle Configuration is attached to.

newCreateStudioLifecycleConfig

Arguments

:: Text
$sel:studioLifecycleConfigName:CreateStudioLifecycleConfig'

-> Text
$sel:studioLifecycleConfigContent:CreateStudioLifecycleConfig'

-> StudioLifecycleConfigAppType
$sel:studioLifecycleConfigAppType:CreateStudioLifecycleConfig'

-> CreateStudioLifecycleConfig

Create a value of CreateStudioLifecycleConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateStudioLifecycleConfig', createStudioLifecycleConfig_tags - Tags to be associated with the Lifecycle Configuration. Each tag consists of a key and an optional value. Tag keys must be unique per resource. Tags are searchable using the Search API.

$sel:studioLifecycleConfigName:CreateStudioLifecycleConfig', createStudioLifecycleConfig_studioLifecycleConfigName - The name of the Studio Lifecycle Configuration to create.

$sel:studioLifecycleConfigContent:CreateStudioLifecycleConfig', createStudioLifecycleConfig_studioLifecycleConfigContent - The content of your Studio Lifecycle Configuration script. This content must be base64 encoded.

$sel:studioLifecycleConfigAppType:CreateStudioLifecycleConfig', createStudioLifecycleConfig_studioLifecycleConfigAppType - The App type that the Lifecycle Configuration is attached to.

Request Lenses

createStudioLifecycleConfig_tags :: Lens' CreateStudioLifecycleConfig (Maybe [Tag])

Tags to be associated with the Lifecycle Configuration. Each tag consists of a key and an optional value. Tag keys must be unique per resource. Tags are searchable using the Search API.

createStudioLifecycleConfig_studioLifecycleConfigName :: Lens' CreateStudioLifecycleConfig Text

The name of the Studio Lifecycle Configuration to create.

createStudioLifecycleConfig_studioLifecycleConfigContent :: Lens' CreateStudioLifecycleConfig Text

The content of your Studio Lifecycle Configuration script. This content must be base64 encoded.

createStudioLifecycleConfig_studioLifecycleConfigAppType :: Lens' CreateStudioLifecycleConfig StudioLifecycleConfigAppType

The App type that the Lifecycle Configuration is attached to.

Destructuring the Response

data CreateStudioLifecycleConfigResponse

See: newCreateStudioLifecycleConfigResponse smart constructor.

Constructors

CreateStudioLifecycleConfigResponse'

Fields

studioLifecycleConfigArn :: Maybe Text

The ARN of your created Lifecycle Configuration.

httpStatus :: Int

The response's http status code.

newCreateStudioLifecycleConfigResponse

Arguments

:: Int
$sel:httpStatus:CreateStudioLifecycleConfigResponse'

-> CreateStudioLifecycleConfigResponse

Create a value of CreateStudioLifecycleConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioLifecycleConfigArn:CreateStudioLifecycleConfigResponse', createStudioLifecycleConfigResponse_studioLifecycleConfigArn - The ARN of your created Lifecycle Configuration.

$sel:httpStatus:CreateStudioLifecycleConfigResponse', createStudioLifecycleConfigResponse_httpStatus - The response's http status code.

Response Lenses

createStudioLifecycleConfigResponse_studioLifecycleConfigArn :: Lens' CreateStudioLifecycleConfigResponse (Maybe Text)

The ARN of your created Lifecycle Configuration.

createStudioLifecycleConfigResponse_httpStatus :: Lens' CreateStudioLifecycleConfigResponse Int

The response's http status code.