Amazonka.SageMaker.DeleteModelExplainabilityJobDefinition

Description

Deletes an Amazon SageMaker model explainability job definition.

Creating a Request

data DeleteModelExplainabilityJobDefinition

See: newDeleteModelExplainabilityJobDefinition smart constructor.

Constructors

DeleteModelExplainabilityJobDefinition'

Fields

jobDefinitionName :: Text

The name of the model explainability job definition to delete.

newDeleteModelExplainabilityJobDefinition

Arguments

:: Text
$sel:jobDefinitionName:DeleteModelExplainabilityJobDefinition'

-> DeleteModelExplainabilityJobDefinition

Create a value of DeleteModelExplainabilityJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobDefinitionName:DeleteModelExplainabilityJobDefinition', deleteModelExplainabilityJobDefinition_jobDefinitionName - The name of the model explainability job definition to delete.

Request Lenses

deleteModelExplainabilityJobDefinition_jobDefinitionName :: Lens' DeleteModelExplainabilityJobDefinition Text

The name of the model explainability job definition to delete.

Destructuring the Response

data DeleteModelExplainabilityJobDefinitionResponse

See: newDeleteModelExplainabilityJobDefinitionResponse smart constructor.

Constructors

DeleteModelExplainabilityJobDefinitionResponse'

newDeleteModelExplainabilityJobDefinitionResponse :: DeleteModelExplainabilityJobDefinitionResponse

Create a value of DeleteModelExplainabilityJobDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.