Amazonka.SageMaker.DescribeContext

Description

Describes a context.

Creating a Request

data DescribeContext

See: newDescribeContext smart constructor.

Constructors

DescribeContext'

Fields

contextName :: Text

The name of the context to describe.

newDescribeContext

Arguments

:: Text
$sel:contextName:DescribeContext'

-> DescribeContext

Create a value of DescribeContext with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contextName:DescribeContext', describeContext_contextName - The name of the context to describe.

Request Lenses

describeContext_contextName :: Lens' DescribeContext Text

The name of the context to describe.

Destructuring the Response

data DescribeContextResponse

See: newDescribeContextResponse smart constructor.

Constructors

DescribeContextResponse'

Fields

creationTime :: Maybe POSIX

When the context was created.

createdBy :: Maybe UserContext

lastModifiedTime :: Maybe POSIX

When the context was last modified.

contextType :: Maybe Text

The type of the context.

contextArn :: Maybe Text

The Amazon Resource Name (ARN) of the context.

source :: Maybe ContextSource

The source of the context.

contextName :: Maybe Text

The name of the context.

lastModifiedBy :: Maybe UserContext

description :: Maybe Text

The description of the context.

properties :: Maybe (HashMap Text Text)

A list of the context's properties.

httpStatus :: Int

The response's http status code.

newDescribeContextResponse

Arguments

:: Int
$sel:httpStatus:DescribeContextResponse'

-> DescribeContextResponse

Create a value of DescribeContextResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeContextResponse', describeContextResponse_creationTime - When the context was created.

$sel:createdBy:DescribeContextResponse', describeContextResponse_createdBy - Undocumented member.

$sel:lastModifiedTime:DescribeContextResponse', describeContextResponse_lastModifiedTime - When the context was last modified.

$sel:contextType:DescribeContextResponse', describeContextResponse_contextType - The type of the context.

$sel:contextArn:DescribeContextResponse', describeContextResponse_contextArn - The Amazon Resource Name (ARN) of the context.

$sel:source:DescribeContextResponse', describeContextResponse_source - The source of the context.

$sel:contextName:DescribeContext', describeContextResponse_contextName - The name of the context.

$sel:lastModifiedBy:DescribeContextResponse', describeContextResponse_lastModifiedBy - Undocumented member.

$sel:description:DescribeContextResponse', describeContextResponse_description - The description of the context.

$sel:properties:DescribeContextResponse', describeContextResponse_properties - A list of the context's properties.

$sel:httpStatus:DescribeContextResponse', describeContextResponse_httpStatus - The response's http status code.

Response Lenses

describeContextResponse_creationTime :: Lens' DescribeContextResponse (Maybe UTCTime)

When the context was created.

describeContextResponse_createdBy :: Lens' DescribeContextResponse (Maybe UserContext)

Undocumented member.

describeContextResponse_lastModifiedTime :: Lens' DescribeContextResponse (Maybe UTCTime)

When the context was last modified.

describeContextResponse_contextType :: Lens' DescribeContextResponse (Maybe Text)

The type of the context.

describeContextResponse_contextArn :: Lens' DescribeContextResponse (Maybe Text)

The Amazon Resource Name (ARN) of the context.

describeContextResponse_source :: Lens' DescribeContextResponse (Maybe ContextSource)

The source of the context.

describeContextResponse_contextName :: Lens' DescribeContextResponse (Maybe Text)

The name of the context.

describeContextResponse_lastModifiedBy :: Lens' DescribeContextResponse (Maybe UserContext)

Undocumented member.

describeContextResponse_description :: Lens' DescribeContextResponse (Maybe Text)

The description of the context.

describeContextResponse_properties :: Lens' DescribeContextResponse (Maybe (HashMap Text Text))

A list of the context's properties.

describeContextResponse_httpStatus :: Lens' DescribeContextResponse Int

The response's http status code.