Amazonka.SageMaker.DescribeDeviceFleet

Description

A description of the fleet the device belongs to.

Creating a Request

data DescribeDeviceFleet

See: newDescribeDeviceFleet smart constructor.

Constructors

DescribeDeviceFleet'

Fields

deviceFleetName :: Text

The name of the fleet.

newDescribeDeviceFleet

Arguments

:: Text
$sel:deviceFleetName:DescribeDeviceFleet'

-> DescribeDeviceFleet

Create a value of DescribeDeviceFleet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceFleetName:DescribeDeviceFleet', describeDeviceFleet_deviceFleetName - The name of the fleet.

Request Lenses

describeDeviceFleet_deviceFleetName :: Lens' DescribeDeviceFleet Text

The name of the fleet.

Destructuring the Response

data DescribeDeviceFleetResponse

See: newDescribeDeviceFleetResponse smart constructor.

Constructors

DescribeDeviceFleetResponse'

Fields

iotRoleAlias :: Maybe Text

The Amazon Resource Name (ARN) alias created in Amazon Web Services Internet of Things (IoT).

description :: Maybe Text

A description of the fleet.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) that has access to Amazon Web Services Internet of Things (IoT).

httpStatus :: Int

The response's http status code.

deviceFleetName :: Text

The name of the fleet.

deviceFleetArn :: Text

The The Amazon Resource Name (ARN) of the fleet.

outputConfig :: EdgeOutputConfig

The output configuration for storing sampled data.

creationTime :: POSIX

Timestamp of when the device fleet was created.

lastModifiedTime :: POSIX

Timestamp of when the device fleet was last updated.

newDescribeDeviceFleetResponse

Arguments

:: Int
$sel:httpStatus:DescribeDeviceFleetResponse'

-> Text
$sel:deviceFleetName:DescribeDeviceFleet'

-> Text
$sel:deviceFleetArn:DescribeDeviceFleetResponse'

-> EdgeOutputConfig
$sel:outputConfig:DescribeDeviceFleetResponse'

-> UTCTime
$sel:creationTime:DescribeDeviceFleetResponse'

-> UTCTime
$sel:lastModifiedTime:DescribeDeviceFleetResponse'

-> DescribeDeviceFleetResponse

Create a value of DescribeDeviceFleetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iotRoleAlias:DescribeDeviceFleetResponse', describeDeviceFleetResponse_iotRoleAlias - The Amazon Resource Name (ARN) alias created in Amazon Web Services Internet of Things (IoT).

$sel:description:DescribeDeviceFleetResponse', describeDeviceFleetResponse_description - A description of the fleet.

$sel:roleArn:DescribeDeviceFleetResponse', describeDeviceFleetResponse_roleArn - The Amazon Resource Name (ARN) that has access to Amazon Web Services Internet of Things (IoT).

$sel:httpStatus:DescribeDeviceFleetResponse', describeDeviceFleetResponse_httpStatus - The response's http status code.

$sel:deviceFleetName:DescribeDeviceFleet', describeDeviceFleetResponse_deviceFleetName - The name of the fleet.

$sel:deviceFleetArn:DescribeDeviceFleetResponse', describeDeviceFleetResponse_deviceFleetArn - The The Amazon Resource Name (ARN) of the fleet.

$sel:outputConfig:DescribeDeviceFleetResponse', describeDeviceFleetResponse_outputConfig - The output configuration for storing sampled data.

$sel:creationTime:DescribeDeviceFleetResponse', describeDeviceFleetResponse_creationTime - Timestamp of when the device fleet was created.

$sel:lastModifiedTime:DescribeDeviceFleetResponse', describeDeviceFleetResponse_lastModifiedTime - Timestamp of when the device fleet was last updated.

Response Lenses

describeDeviceFleetResponse_iotRoleAlias :: Lens' DescribeDeviceFleetResponse (Maybe Text)

The Amazon Resource Name (ARN) alias created in Amazon Web Services Internet of Things (IoT).

describeDeviceFleetResponse_description :: Lens' DescribeDeviceFleetResponse (Maybe Text)

A description of the fleet.

describeDeviceFleetResponse_roleArn :: Lens' DescribeDeviceFleetResponse (Maybe Text)

The Amazon Resource Name (ARN) that has access to Amazon Web Services Internet of Things (IoT).

describeDeviceFleetResponse_httpStatus :: Lens' DescribeDeviceFleetResponse Int

The response's http status code.

describeDeviceFleetResponse_deviceFleetName :: Lens' DescribeDeviceFleetResponse Text

The name of the fleet.

describeDeviceFleetResponse_deviceFleetArn :: Lens' DescribeDeviceFleetResponse Text

The The Amazon Resource Name (ARN) of the fleet.

describeDeviceFleetResponse_outputConfig :: Lens' DescribeDeviceFleetResponse EdgeOutputConfig

The output configuration for storing sampled data.

describeDeviceFleetResponse_creationTime :: Lens' DescribeDeviceFleetResponse UTCTime

Timestamp of when the device fleet was created.

describeDeviceFleetResponse_lastModifiedTime :: Lens' DescribeDeviceFleetResponse UTCTime

Timestamp of when the device fleet was last updated.