Amazonka.SageMaker.DescribeFlowDefinition

Description

Returns information about the specified flow definition.

Creating a Request

data DescribeFlowDefinition

See: newDescribeFlowDefinition smart constructor.

Constructors

DescribeFlowDefinition'

Fields

flowDefinitionName :: Text

The name of the flow definition.

newDescribeFlowDefinition

Arguments

:: Text
$sel:flowDefinitionName:DescribeFlowDefinition'

-> DescribeFlowDefinition

Create a value of DescribeFlowDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowDefinitionName:DescribeFlowDefinition', describeFlowDefinition_flowDefinitionName - The name of the flow definition.

Request Lenses

describeFlowDefinition_flowDefinitionName :: Lens' DescribeFlowDefinition Text

The name of the flow definition.

Destructuring the Response

data DescribeFlowDefinitionResponse

See: newDescribeFlowDefinitionResponse smart constructor.

Constructors

DescribeFlowDefinitionResponse'

Fields

failureReason :: Maybe Text

The reason your flow definition failed.

humanLoopRequestSource :: Maybe HumanLoopRequestSource

Container for configuring the source of human task requests. Used to specify if Amazon Rekognition or Amazon Textract is used as an integration source.

humanLoopActivationConfig :: Maybe HumanLoopActivationConfig

An object containing information about what triggers a human review workflow.

httpStatus :: Int

The response's http status code.

flowDefinitionArn :: Text

The Amazon Resource Name (ARN) of the flow defintion.

flowDefinitionName :: Text

The Amazon Resource Name (ARN) of the flow definition.

flowDefinitionStatus :: FlowDefinitionStatus

The status of the flow definition. Valid values are listed below.

creationTime :: POSIX

The timestamp when the flow definition was created.

humanLoopConfig :: HumanLoopConfig

An object containing information about who works on the task, the workforce task price, and other task details.

outputConfig :: FlowDefinitionOutputConfig

An object containing information about the output file.

roleArn :: Text

The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) execution role for the flow definition.

newDescribeFlowDefinitionResponse

Arguments

:: Int
$sel:httpStatus:DescribeFlowDefinitionResponse'

-> Text
$sel:flowDefinitionArn:DescribeFlowDefinitionResponse'

-> Text
$sel:flowDefinitionName:DescribeFlowDefinition'

-> FlowDefinitionStatus
$sel:flowDefinitionStatus:DescribeFlowDefinitionResponse'

-> UTCTime
$sel:creationTime:DescribeFlowDefinitionResponse'

-> HumanLoopConfig
$sel:humanLoopConfig:DescribeFlowDefinitionResponse'

-> FlowDefinitionOutputConfig
$sel:outputConfig:DescribeFlowDefinitionResponse'

-> Text
$sel:roleArn:DescribeFlowDefinitionResponse'

-> DescribeFlowDefinitionResponse

Create a value of DescribeFlowDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DescribeFlowDefinitionResponse', describeFlowDefinitionResponse_failureReason - The reason your flow definition failed.

$sel:humanLoopRequestSource:DescribeFlowDefinitionResponse', describeFlowDefinitionResponse_humanLoopRequestSource - Container for configuring the source of human task requests. Used to specify if Amazon Rekognition or Amazon Textract is used as an integration source.

$sel:humanLoopActivationConfig:DescribeFlowDefinitionResponse', describeFlowDefinitionResponse_humanLoopActivationConfig - An object containing information about what triggers a human review workflow.

$sel:httpStatus:DescribeFlowDefinitionResponse', describeFlowDefinitionResponse_httpStatus - The response's http status code.

$sel:flowDefinitionArn:DescribeFlowDefinitionResponse', describeFlowDefinitionResponse_flowDefinitionArn - The Amazon Resource Name (ARN) of the flow defintion.

$sel:flowDefinitionName:DescribeFlowDefinition', describeFlowDefinitionResponse_flowDefinitionName - The Amazon Resource Name (ARN) of the flow definition.

$sel:flowDefinitionStatus:DescribeFlowDefinitionResponse', describeFlowDefinitionResponse_flowDefinitionStatus - The status of the flow definition. Valid values are listed below.

$sel:creationTime:DescribeFlowDefinitionResponse', describeFlowDefinitionResponse_creationTime - The timestamp when the flow definition was created.

$sel:humanLoopConfig:DescribeFlowDefinitionResponse', describeFlowDefinitionResponse_humanLoopConfig - An object containing information about who works on the task, the workforce task price, and other task details.

$sel:outputConfig:DescribeFlowDefinitionResponse', describeFlowDefinitionResponse_outputConfig - An object containing information about the output file.

$sel:roleArn:DescribeFlowDefinitionResponse', describeFlowDefinitionResponse_roleArn - The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) execution role for the flow definition.

Response Lenses

describeFlowDefinitionResponse_failureReason :: Lens' DescribeFlowDefinitionResponse (Maybe Text)

The reason your flow definition failed.

describeFlowDefinitionResponse_humanLoopRequestSource :: Lens' DescribeFlowDefinitionResponse (Maybe HumanLoopRequestSource)

Container for configuring the source of human task requests. Used to specify if Amazon Rekognition or Amazon Textract is used as an integration source.

describeFlowDefinitionResponse_humanLoopActivationConfig :: Lens' DescribeFlowDefinitionResponse (Maybe HumanLoopActivationConfig)

An object containing information about what triggers a human review workflow.

describeFlowDefinitionResponse_httpStatus :: Lens' DescribeFlowDefinitionResponse Int

The response's http status code.

describeFlowDefinitionResponse_flowDefinitionArn :: Lens' DescribeFlowDefinitionResponse Text

The Amazon Resource Name (ARN) of the flow defintion.

describeFlowDefinitionResponse_flowDefinitionName :: Lens' DescribeFlowDefinitionResponse Text

The Amazon Resource Name (ARN) of the flow definition.

describeFlowDefinitionResponse_flowDefinitionStatus :: Lens' DescribeFlowDefinitionResponse FlowDefinitionStatus

The status of the flow definition. Valid values are listed below.

describeFlowDefinitionResponse_creationTime :: Lens' DescribeFlowDefinitionResponse UTCTime

The timestamp when the flow definition was created.

describeFlowDefinitionResponse_humanLoopConfig :: Lens' DescribeFlowDefinitionResponse HumanLoopConfig

An object containing information about who works on the task, the workforce task price, and other task details.

describeFlowDefinitionResponse_outputConfig :: Lens' DescribeFlowDefinitionResponse FlowDefinitionOutputConfig

An object containing information about the output file.

describeFlowDefinitionResponse_roleArn :: Lens' DescribeFlowDefinitionResponse Text

The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) execution role for the flow definition.