Amazonka.SageMaker.DescribeImageVersion

Description

Describes a version of a SageMaker image.

Creating a Request

data DescribeImageVersion

See: newDescribeImageVersion smart constructor.

Constructors

DescribeImageVersion'

Fields

version :: Maybe Natural

The version of the image. If not specified, the latest version is described.

imageName :: Text

The name of the image.

newDescribeImageVersion

Arguments

:: Text
$sel:imageName:DescribeImageVersion'

-> DescribeImageVersion

Create a value of DescribeImageVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:DescribeImageVersion', describeImageVersion_version - The version of the image. If not specified, the latest version is described.

$sel:imageName:DescribeImageVersion', describeImageVersion_imageName - The name of the image.

Request Lenses

describeImageVersion_version :: Lens' DescribeImageVersion (Maybe Natural)

The version of the image. If not specified, the latest version is described.

describeImageVersion_imageName :: Lens' DescribeImageVersion Text

The name of the image.

Destructuring the Response

data DescribeImageVersionResponse

See: newDescribeImageVersionResponse smart constructor.

Constructors

DescribeImageVersionResponse'

Fields

creationTime :: Maybe POSIX

When the version was created.

failureReason :: Maybe Text

When a create or delete operation fails, the reason for the failure.

containerImage :: Maybe Text

The registry path of the container image that contains this image version.

lastModifiedTime :: Maybe POSIX

When the version was last modified.

imageVersionStatus :: Maybe ImageVersionStatus

The status of the version.

version :: Maybe Natural

The version number.

baseImage :: Maybe Text

The registry path of the container image on which this image version is based.

imageArn :: Maybe Text

The Amazon Resource Name (ARN) of the image the version is based on.

imageVersionArn :: Maybe Text

The ARN of the version.

httpStatus :: Int

The response's http status code.

newDescribeImageVersionResponse

Arguments

:: Int
$sel:httpStatus:DescribeImageVersionResponse'

-> DescribeImageVersionResponse

Create a value of DescribeImageVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeImageVersionResponse', describeImageVersionResponse_creationTime - When the version was created.

$sel:failureReason:DescribeImageVersionResponse', describeImageVersionResponse_failureReason - When a create or delete operation fails, the reason for the failure.

$sel:containerImage:DescribeImageVersionResponse', describeImageVersionResponse_containerImage - The registry path of the container image that contains this image version.

$sel:lastModifiedTime:DescribeImageVersionResponse', describeImageVersionResponse_lastModifiedTime - When the version was last modified.

$sel:imageVersionStatus:DescribeImageVersionResponse', describeImageVersionResponse_imageVersionStatus - The status of the version.

$sel:version:DescribeImageVersion', describeImageVersionResponse_version - The version number.

$sel:baseImage:DescribeImageVersionResponse', describeImageVersionResponse_baseImage - The registry path of the container image on which this image version is based.

$sel:imageArn:DescribeImageVersionResponse', describeImageVersionResponse_imageArn - The Amazon Resource Name (ARN) of the image the version is based on.

$sel:imageVersionArn:DescribeImageVersionResponse', describeImageVersionResponse_imageVersionArn - The ARN of the version.

$sel:httpStatus:DescribeImageVersionResponse', describeImageVersionResponse_httpStatus - The response's http status code.

Response Lenses

describeImageVersionResponse_creationTime :: Lens' DescribeImageVersionResponse (Maybe UTCTime)

When the version was created.

describeImageVersionResponse_failureReason :: Lens' DescribeImageVersionResponse (Maybe Text)

When a create or delete operation fails, the reason for the failure.

describeImageVersionResponse_containerImage :: Lens' DescribeImageVersionResponse (Maybe Text)

The registry path of the container image that contains this image version.

describeImageVersionResponse_lastModifiedTime :: Lens' DescribeImageVersionResponse (Maybe UTCTime)

When the version was last modified.

describeImageVersionResponse_imageVersionStatus :: Lens' DescribeImageVersionResponse (Maybe ImageVersionStatus)

The status of the version.

describeImageVersionResponse_version :: Lens' DescribeImageVersionResponse (Maybe Natural)

The version number.

describeImageVersionResponse_baseImage :: Lens' DescribeImageVersionResponse (Maybe Text)

The registry path of the container image on which this image version is based.

describeImageVersionResponse_imageArn :: Lens' DescribeImageVersionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the image the version is based on.

describeImageVersionResponse_imageVersionArn :: Lens' DescribeImageVersionResponse (Maybe Text)

The ARN of the version.

describeImageVersionResponse_httpStatus :: Lens' DescribeImageVersionResponse Int

The response's http status code.