Amazonka.SageMaker.DescribeLabelingJob

Description

Gets information about a labeling job.

Creating a Request

data DescribeLabelingJob

See: newDescribeLabelingJob smart constructor.

Constructors

DescribeLabelingJob'

Fields

labelingJobName :: Text

The name of the labeling job to return information for.

newDescribeLabelingJob

Arguments

:: Text
$sel:labelingJobName:DescribeLabelingJob'

-> DescribeLabelingJob

Create a value of DescribeLabelingJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:labelingJobName:DescribeLabelingJob', describeLabelingJob_labelingJobName - The name of the labeling job to return information for.

Request Lenses

describeLabelingJob_labelingJobName :: Lens' DescribeLabelingJob Text

The name of the labeling job to return information for.

Destructuring the Response

data DescribeLabelingJobResponse

See: newDescribeLabelingJobResponse smart constructor.

Constructors

DescribeLabelingJobResponse'

Fields

failureReason :: Maybe Text

If the job failed, the reason that it failed.

labelingJobAlgorithmsConfig :: Maybe LabelingJobAlgorithmsConfig

Configuration information for automated data labeling.

labelCategoryConfigS3Uri :: Maybe Text

The S3 location of the JSON file that defines the categories used to label data objects. Please note the following label-category limits:

Semantic segmentation labeling jobs using automated labeling: 20 labels

Box bounding labeling jobs (all): 10 labels

The file is a JSON structure in the following format:

{

 "document-version": "2018-11-28"

 "labels": [

 {

 "label": "label 1"

 },

 {

 "label": "label 2"

 },

 ...

 {

 "label": "label n"

 }

 ]

}

stoppingConditions :: Maybe LabelingJobStoppingConditions

A set of conditions for stopping a labeling job. If any of the conditions are met, the job is automatically stopped.

labelAttributeName :: Maybe Text

The attribute used as the label in the output manifest file.

labelingJobOutput :: Maybe LabelingJobOutput

The location of the output produced by the labeling job.

tags :: Maybe [Tag]

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

httpStatus :: Int

The response's http status code.

labelingJobStatus :: LabelingJobStatus

The processing status of the labeling job.

labelCounters :: LabelCounters

Provides a breakdown of the number of data objects labeled by humans, the number of objects labeled by machine, the number of objects than couldn't be labeled, and the total number of objects labeled.

creationTime :: POSIX

The date and time that the labeling job was created.

lastModifiedTime :: POSIX

The date and time that the labeling job was last updated.

jobReferenceCode :: Text

A unique identifier for work done as part of a labeling job.

labelingJobName :: Text

The name assigned to the labeling job when it was created.

labelingJobArn :: Text

The Amazon Resource Name (ARN) of the labeling job.

inputConfig :: LabelingJobInputConfig

Input configuration information for the labeling job, such as the Amazon S3 location of the data objects and the location of the manifest file that describes the data objects.

outputConfig :: LabelingJobOutputConfig

The location of the job's output data and the Amazon Web Services Key Management Service key ID for the key used to encrypt the output data, if any.

roleArn :: Text

The Amazon Resource Name (ARN) that Amazon SageMaker assumes to perform tasks on your behalf during data labeling.

humanTaskConfig :: HumanTaskConfig

Configuration information required for human workers to complete a labeling task.

newDescribeLabelingJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeLabelingJobResponse'

-> LabelingJobStatus
$sel:labelingJobStatus:DescribeLabelingJobResponse'

-> LabelCounters
$sel:labelCounters:DescribeLabelingJobResponse'

-> UTCTime
$sel:creationTime:DescribeLabelingJobResponse'

-> UTCTime
$sel:lastModifiedTime:DescribeLabelingJobResponse'

-> Text
$sel:jobReferenceCode:DescribeLabelingJobResponse'

-> Text
$sel:labelingJobName:DescribeLabelingJob'

-> Text
$sel:labelingJobArn:DescribeLabelingJobResponse'

-> LabelingJobInputConfig
$sel:inputConfig:DescribeLabelingJobResponse'

-> LabelingJobOutputConfig
$sel:outputConfig:DescribeLabelingJobResponse'

-> Text
$sel:roleArn:DescribeLabelingJobResponse'

-> HumanTaskConfig
$sel:humanTaskConfig:DescribeLabelingJobResponse'

-> DescribeLabelingJobResponse

Create a value of DescribeLabelingJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DescribeLabelingJobResponse', describeLabelingJobResponse_failureReason - If the job failed, the reason that it failed.

$sel:labelingJobAlgorithmsConfig:DescribeLabelingJobResponse', describeLabelingJobResponse_labelingJobAlgorithmsConfig - Configuration information for automated data labeling.

$sel:labelCategoryConfigS3Uri:DescribeLabelingJobResponse', describeLabelingJobResponse_labelCategoryConfigS3Uri - The S3 location of the JSON file that defines the categories used to label data objects. Please note the following label-category limits:

Semantic segmentation labeling jobs using automated labeling: 20 labels

Box bounding labeling jobs (all): 10 labels

The file is a JSON structure in the following format:

{

 "document-version": "2018-11-28"

 "labels": [

 {

 "label": "label 1"

 },

 {

 "label": "label 2"

 },

 ...

 {

 "label": "label n"

 }

 ]

}

$sel:stoppingConditions:DescribeLabelingJobResponse', describeLabelingJobResponse_stoppingConditions - A set of conditions for stopping a labeling job. If any of the conditions are met, the job is automatically stopped.

$sel:labelAttributeName:DescribeLabelingJobResponse', describeLabelingJobResponse_labelAttributeName - The attribute used as the label in the output manifest file.

$sel:labelingJobOutput:DescribeLabelingJobResponse', describeLabelingJobResponse_labelingJobOutput - The location of the output produced by the labeling job.

$sel:tags:DescribeLabelingJobResponse', describeLabelingJobResponse_tags - An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

$sel:httpStatus:DescribeLabelingJobResponse', describeLabelingJobResponse_httpStatus - The response's http status code.

$sel:labelingJobStatus:DescribeLabelingJobResponse', describeLabelingJobResponse_labelingJobStatus - The processing status of the labeling job.

$sel:labelCounters:DescribeLabelingJobResponse', describeLabelingJobResponse_labelCounters - Provides a breakdown of the number of data objects labeled by humans, the number of objects labeled by machine, the number of objects than couldn't be labeled, and the total number of objects labeled.

$sel:creationTime:DescribeLabelingJobResponse', describeLabelingJobResponse_creationTime - The date and time that the labeling job was created.

$sel:lastModifiedTime:DescribeLabelingJobResponse', describeLabelingJobResponse_lastModifiedTime - The date and time that the labeling job was last updated.

$sel:jobReferenceCode:DescribeLabelingJobResponse', describeLabelingJobResponse_jobReferenceCode - A unique identifier for work done as part of a labeling job.

$sel:labelingJobName:DescribeLabelingJob', describeLabelingJobResponse_labelingJobName - The name assigned to the labeling job when it was created.

$sel:labelingJobArn:DescribeLabelingJobResponse', describeLabelingJobResponse_labelingJobArn - The Amazon Resource Name (ARN) of the labeling job.

$sel:inputConfig:DescribeLabelingJobResponse', describeLabelingJobResponse_inputConfig - Input configuration information for the labeling job, such as the Amazon S3 location of the data objects and the location of the manifest file that describes the data objects.

$sel:outputConfig:DescribeLabelingJobResponse', describeLabelingJobResponse_outputConfig - The location of the job's output data and the Amazon Web Services Key Management Service key ID for the key used to encrypt the output data, if any.

$sel:roleArn:DescribeLabelingJobResponse', describeLabelingJobResponse_roleArn - The Amazon Resource Name (ARN) that Amazon SageMaker assumes to perform tasks on your behalf during data labeling.

$sel:humanTaskConfig:DescribeLabelingJobResponse', describeLabelingJobResponse_humanTaskConfig - Configuration information required for human workers to complete a labeling task.

Response Lenses

describeLabelingJobResponse_failureReason :: Lens' DescribeLabelingJobResponse (Maybe Text)

If the job failed, the reason that it failed.

describeLabelingJobResponse_labelingJobAlgorithmsConfig :: Lens' DescribeLabelingJobResponse (Maybe LabelingJobAlgorithmsConfig)

Configuration information for automated data labeling.

describeLabelingJobResponse_labelCategoryConfigS3Uri :: Lens' DescribeLabelingJobResponse (Maybe Text)

The S3 location of the JSON file that defines the categories used to label data objects. Please note the following label-category limits:

Semantic segmentation labeling jobs using automated labeling: 20 labels

Box bounding labeling jobs (all): 10 labels

The file is a JSON structure in the following format:

{

 "document-version": "2018-11-28"

 "labels": [

 {

 "label": "label 1"

 },

 {

 "label": "label 2"

 },

 ...

 {

 "label": "label n"

 }

 ]

}

describeLabelingJobResponse_stoppingConditions :: Lens' DescribeLabelingJobResponse (Maybe LabelingJobStoppingConditions)

A set of conditions for stopping a labeling job. If any of the conditions are met, the job is automatically stopped.

describeLabelingJobResponse_labelAttributeName :: Lens' DescribeLabelingJobResponse (Maybe Text)

The attribute used as the label in the output manifest file.

describeLabelingJobResponse_labelingJobOutput :: Lens' DescribeLabelingJobResponse (Maybe LabelingJobOutput)

The location of the output produced by the labeling job.

describeLabelingJobResponse_tags :: Lens' DescribeLabelingJobResponse (Maybe [Tag])

An array of key-value pairs. You can use tags to categorize your Amazon Web Services resources in different ways, for example, by purpose, owner, or environment. For more information, see Tagging Amazon Web Services Resources.

describeLabelingJobResponse_httpStatus :: Lens' DescribeLabelingJobResponse Int

The response's http status code.

describeLabelingJobResponse_labelingJobStatus :: Lens' DescribeLabelingJobResponse LabelingJobStatus

The processing status of the labeling job.

describeLabelingJobResponse_labelCounters :: Lens' DescribeLabelingJobResponse LabelCounters

Provides a breakdown of the number of data objects labeled by humans, the number of objects labeled by machine, the number of objects than couldn't be labeled, and the total number of objects labeled.

describeLabelingJobResponse_creationTime :: Lens' DescribeLabelingJobResponse UTCTime

The date and time that the labeling job was created.

describeLabelingJobResponse_lastModifiedTime :: Lens' DescribeLabelingJobResponse UTCTime

The date and time that the labeling job was last updated.

describeLabelingJobResponse_jobReferenceCode :: Lens' DescribeLabelingJobResponse Text

A unique identifier for work done as part of a labeling job.

describeLabelingJobResponse_labelingJobName :: Lens' DescribeLabelingJobResponse Text

The name assigned to the labeling job when it was created.

describeLabelingJobResponse_labelingJobArn :: Lens' DescribeLabelingJobResponse Text

The Amazon Resource Name (ARN) of the labeling job.

describeLabelingJobResponse_inputConfig :: Lens' DescribeLabelingJobResponse LabelingJobInputConfig

Input configuration information for the labeling job, such as the Amazon S3 location of the data objects and the location of the manifest file that describes the data objects.

describeLabelingJobResponse_outputConfig :: Lens' DescribeLabelingJobResponse LabelingJobOutputConfig

The location of the job's output data and the Amazon Web Services Key Management Service key ID for the key used to encrypt the output data, if any.

describeLabelingJobResponse_roleArn :: Lens' DescribeLabelingJobResponse Text

The Amazon Resource Name (ARN) that Amazon SageMaker assumes to perform tasks on your behalf during data labeling.

describeLabelingJobResponse_humanTaskConfig :: Lens' DescribeLabelingJobResponse HumanTaskConfig

Configuration information required for human workers to complete a labeling task.