Amazonka.SageMaker.DescribeModelPackage

Description

Returns a description of the specified model package, which is used to create Amazon SageMaker models or list them on Amazon Web Services Marketplace.

To create models in Amazon SageMaker, buyers can subscribe to model packages listed on Amazon Web Services Marketplace.

Creating a Request

data DescribeModelPackage

See: newDescribeModelPackage smart constructor.

Constructors

DescribeModelPackage'

Fields

modelPackageName :: Text

The name or Amazon Resource Name (ARN) of the model package to describe.

When you specify a name, the name must have 1 to 63 characters. Valid characters are a-z, A-Z, 0-9, and - (hyphen).

newDescribeModelPackage

Arguments

:: Text
$sel:modelPackageName:DescribeModelPackage'

-> DescribeModelPackage

Create a value of DescribeModelPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelPackageName:DescribeModelPackage', describeModelPackage_modelPackageName - The name or Amazon Resource Name (ARN) of the model package to describe.

When you specify a name, the name must have 1 to 63 characters. Valid characters are a-z, A-Z, 0-9, and - (hyphen).

Request Lenses

describeModelPackage_modelPackageName :: Lens' DescribeModelPackage Text

The name or Amazon Resource Name (ARN) of the model package to describe.

When you specify a name, the name must have 1 to 63 characters. Valid characters are a-z, A-Z, 0-9, and - (hyphen).

Destructuring the Response

data DescribeModelPackageResponse

See: newDescribeModelPackageResponse smart constructor.

Constructors

DescribeModelPackageResponse'

Fields

metadataProperties :: Maybe MetadataProperties

modelApprovalStatus :: Maybe ModelApprovalStatus

The approval status of the model package.

sourceAlgorithmSpecification :: Maybe SourceAlgorithmSpecification

Details about the algorithm that was used to create the model package.

modelMetrics :: Maybe ModelMetrics

Metrics for the model.

modelPackageDescription :: Maybe Text

A brief summary of the model package.

createdBy :: Maybe UserContext

lastModifiedTime :: Maybe POSIX

The last time the model package was modified.

validationSpecification :: Maybe ModelPackageValidationSpecification

Configurations for one or more transform jobs that Amazon SageMaker runs to test the model package.

inferenceSpecification :: Maybe InferenceSpecification

Details about inference jobs that can be run with models based on this model package.

approvalDescription :: Maybe Text

A description provided for the model approval.

modelPackageVersion :: Maybe Natural

The version of the model package.

certifyForMarketplace :: Maybe Bool

Whether the model package is certified for listing on Amazon Web Services Marketplace.

modelPackageGroupName :: Maybe Text

If the model is a versioned model, the name of the model group that the versioned model belongs to.

lastModifiedBy :: Maybe UserContext

httpStatus :: Int

The response's http status code.

modelPackageName :: Text

The name of the model package being described.

modelPackageArn :: Text

The Amazon Resource Name (ARN) of the model package.

creationTime :: POSIX

A timestamp specifying when the model package was created.

modelPackageStatus :: ModelPackageStatus

The current status of the model package.

modelPackageStatusDetails :: ModelPackageStatusDetails

Details about the current status of the model package.

newDescribeModelPackageResponse

Arguments

:: Int
$sel:httpStatus:DescribeModelPackageResponse'

-> Text
$sel:modelPackageName:DescribeModelPackage'

-> Text
$sel:modelPackageArn:DescribeModelPackageResponse'

-> UTCTime
$sel:creationTime:DescribeModelPackageResponse'

-> ModelPackageStatus
$sel:modelPackageStatus:DescribeModelPackageResponse'

-> ModelPackageStatusDetails
$sel:modelPackageStatusDetails:DescribeModelPackageResponse'

-> DescribeModelPackageResponse

Create a value of DescribeModelPackageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadataProperties:DescribeModelPackageResponse', describeModelPackageResponse_metadataProperties - Undocumented member.

$sel:modelApprovalStatus:DescribeModelPackageResponse', describeModelPackageResponse_modelApprovalStatus - The approval status of the model package.

$sel:sourceAlgorithmSpecification:DescribeModelPackageResponse', describeModelPackageResponse_sourceAlgorithmSpecification - Details about the algorithm that was used to create the model package.

$sel:modelMetrics:DescribeModelPackageResponse', describeModelPackageResponse_modelMetrics - Metrics for the model.

$sel:modelPackageDescription:DescribeModelPackageResponse', describeModelPackageResponse_modelPackageDescription - A brief summary of the model package.

$sel:createdBy:DescribeModelPackageResponse', describeModelPackageResponse_createdBy - Undocumented member.

$sel:lastModifiedTime:DescribeModelPackageResponse', describeModelPackageResponse_lastModifiedTime - The last time the model package was modified.

$sel:validationSpecification:DescribeModelPackageResponse', describeModelPackageResponse_validationSpecification - Configurations for one or more transform jobs that Amazon SageMaker runs to test the model package.

$sel:inferenceSpecification:DescribeModelPackageResponse', describeModelPackageResponse_inferenceSpecification - Details about inference jobs that can be run with models based on this model package.

$sel:approvalDescription:DescribeModelPackageResponse', describeModelPackageResponse_approvalDescription - A description provided for the model approval.

$sel:modelPackageVersion:DescribeModelPackageResponse', describeModelPackageResponse_modelPackageVersion - The version of the model package.

$sel:certifyForMarketplace:DescribeModelPackageResponse', describeModelPackageResponse_certifyForMarketplace - Whether the model package is certified for listing on Amazon Web Services Marketplace.

$sel:modelPackageGroupName:DescribeModelPackageResponse', describeModelPackageResponse_modelPackageGroupName - If the model is a versioned model, the name of the model group that the versioned model belongs to.

$sel:lastModifiedBy:DescribeModelPackageResponse', describeModelPackageResponse_lastModifiedBy - Undocumented member.

$sel:httpStatus:DescribeModelPackageResponse', describeModelPackageResponse_httpStatus - The response's http status code.

$sel:modelPackageName:DescribeModelPackage', describeModelPackageResponse_modelPackageName - The name of the model package being described.

$sel:modelPackageArn:DescribeModelPackageResponse', describeModelPackageResponse_modelPackageArn - The Amazon Resource Name (ARN) of the model package.

$sel:creationTime:DescribeModelPackageResponse', describeModelPackageResponse_creationTime - A timestamp specifying when the model package was created.

$sel:modelPackageStatus:DescribeModelPackageResponse', describeModelPackageResponse_modelPackageStatus - The current status of the model package.

$sel:modelPackageStatusDetails:DescribeModelPackageResponse', describeModelPackageResponse_modelPackageStatusDetails - Details about the current status of the model package.

Response Lenses

describeModelPackageResponse_metadataProperties :: Lens' DescribeModelPackageResponse (Maybe MetadataProperties)

Undocumented member.

describeModelPackageResponse_modelApprovalStatus :: Lens' DescribeModelPackageResponse (Maybe ModelApprovalStatus)

The approval status of the model package.

describeModelPackageResponse_sourceAlgorithmSpecification :: Lens' DescribeModelPackageResponse (Maybe SourceAlgorithmSpecification)

Details about the algorithm that was used to create the model package.

describeModelPackageResponse_modelMetrics :: Lens' DescribeModelPackageResponse (Maybe ModelMetrics)

Metrics for the model.

describeModelPackageResponse_modelPackageDescription :: Lens' DescribeModelPackageResponse (Maybe Text)

A brief summary of the model package.

describeModelPackageResponse_createdBy :: Lens' DescribeModelPackageResponse (Maybe UserContext)

Undocumented member.

describeModelPackageResponse_lastModifiedTime :: Lens' DescribeModelPackageResponse (Maybe UTCTime)

The last time the model package was modified.

describeModelPackageResponse_validationSpecification :: Lens' DescribeModelPackageResponse (Maybe ModelPackageValidationSpecification)

Configurations for one or more transform jobs that Amazon SageMaker runs to test the model package.

describeModelPackageResponse_inferenceSpecification :: Lens' DescribeModelPackageResponse (Maybe InferenceSpecification)

Details about inference jobs that can be run with models based on this model package.

describeModelPackageResponse_approvalDescription :: Lens' DescribeModelPackageResponse (Maybe Text)

A description provided for the model approval.

describeModelPackageResponse_modelPackageVersion :: Lens' DescribeModelPackageResponse (Maybe Natural)

The version of the model package.

describeModelPackageResponse_certifyForMarketplace :: Lens' DescribeModelPackageResponse (Maybe Bool)

Whether the model package is certified for listing on Amazon Web Services Marketplace.

describeModelPackageResponse_modelPackageGroupName :: Lens' DescribeModelPackageResponse (Maybe Text)

If the model is a versioned model, the name of the model group that the versioned model belongs to.

describeModelPackageResponse_lastModifiedBy :: Lens' DescribeModelPackageResponse (Maybe UserContext)

Undocumented member.

describeModelPackageResponse_httpStatus :: Lens' DescribeModelPackageResponse Int

The response's http status code.

describeModelPackageResponse_modelPackageName :: Lens' DescribeModelPackageResponse Text

The name of the model package being described.

describeModelPackageResponse_modelPackageArn :: Lens' DescribeModelPackageResponse Text

The Amazon Resource Name (ARN) of the model package.

describeModelPackageResponse_creationTime :: Lens' DescribeModelPackageResponse UTCTime

A timestamp specifying when the model package was created.

describeModelPackageResponse_modelPackageStatus :: Lens' DescribeModelPackageResponse ModelPackageStatus

The current status of the model package.

describeModelPackageResponse_modelPackageStatusDetails :: Lens' DescribeModelPackageResponse ModelPackageStatusDetails

Details about the current status of the model package.