Amazonka.SageMaker.DescribePipeline

Description

Describes the details of a pipeline.

Creating a Request

data DescribePipeline

See: newDescribePipeline smart constructor.

Constructors

DescribePipeline'

Fields

pipelineName :: Text

The name of the pipeline to describe.

newDescribePipeline

Arguments

:: Text
$sel:pipelineName:DescribePipeline'

-> DescribePipeline

Create a value of DescribePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineName:DescribePipeline', describePipeline_pipelineName - The name of the pipeline to describe.

Request Lenses

describePipeline_pipelineName :: Lens' DescribePipeline Text

The name of the pipeline to describe.

Destructuring the Response

data DescribePipelineResponse

See: newDescribePipelineResponse smart constructor.

Constructors

DescribePipelineResponse'

Fields

creationTime :: Maybe POSIX

The time when the pipeline was created.

pipelineDisplayName :: Maybe Text

The display name of the pipeline.

pipelineName :: Maybe Text

The name of the pipeline.

createdBy :: Maybe UserContext

lastRunTime :: Maybe POSIX

The time when the pipeline was last run.

lastModifiedTime :: Maybe POSIX

The time when the pipeline was last modified.

pipelineStatus :: Maybe PipelineStatus

The status of the pipeline execution.

pipelineDescription :: Maybe Text

The description of the pipeline.

pipelineArn :: Maybe Text

The Amazon Resource Name (ARN) of the pipeline.

pipelineDefinition :: Maybe Text

The JSON pipeline definition.

lastModifiedBy :: Maybe UserContext

roleArn :: Maybe Text

The Amazon Resource Name (ARN) that the pipeline uses to execute.

httpStatus :: Int

The response's http status code.

newDescribePipelineResponse

Arguments

:: Int
$sel:httpStatus:DescribePipelineResponse'

-> DescribePipelineResponse

Create a value of DescribePipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribePipelineResponse', describePipelineResponse_creationTime - The time when the pipeline was created.

$sel:pipelineDisplayName:DescribePipelineResponse', describePipelineResponse_pipelineDisplayName - The display name of the pipeline.

$sel:pipelineName:DescribePipeline', describePipelineResponse_pipelineName - The name of the pipeline.

$sel:createdBy:DescribePipelineResponse', describePipelineResponse_createdBy - Undocumented member.

$sel:lastRunTime:DescribePipelineResponse', describePipelineResponse_lastRunTime - The time when the pipeline was last run.

$sel:lastModifiedTime:DescribePipelineResponse', describePipelineResponse_lastModifiedTime - The time when the pipeline was last modified.

$sel:pipelineStatus:DescribePipelineResponse', describePipelineResponse_pipelineStatus - The status of the pipeline execution.

$sel:pipelineDescription:DescribePipelineResponse', describePipelineResponse_pipelineDescription - The description of the pipeline.

$sel:pipelineArn:DescribePipelineResponse', describePipelineResponse_pipelineArn - The Amazon Resource Name (ARN) of the pipeline.

$sel:pipelineDefinition:DescribePipelineResponse', describePipelineResponse_pipelineDefinition - The JSON pipeline definition.

$sel:lastModifiedBy:DescribePipelineResponse', describePipelineResponse_lastModifiedBy - Undocumented member.

$sel:roleArn:DescribePipelineResponse', describePipelineResponse_roleArn - The Amazon Resource Name (ARN) that the pipeline uses to execute.

$sel:httpStatus:DescribePipelineResponse', describePipelineResponse_httpStatus - The response's http status code.

Response Lenses

describePipelineResponse_creationTime :: Lens' DescribePipelineResponse (Maybe UTCTime)

The time when the pipeline was created.

describePipelineResponse_pipelineDisplayName :: Lens' DescribePipelineResponse (Maybe Text)

The display name of the pipeline.

describePipelineResponse_pipelineName :: Lens' DescribePipelineResponse (Maybe Text)

The name of the pipeline.

describePipelineResponse_createdBy :: Lens' DescribePipelineResponse (Maybe UserContext)

Undocumented member.

describePipelineResponse_lastRunTime :: Lens' DescribePipelineResponse (Maybe UTCTime)

The time when the pipeline was last run.

describePipelineResponse_lastModifiedTime :: Lens' DescribePipelineResponse (Maybe UTCTime)

The time when the pipeline was last modified.

describePipelineResponse_pipelineStatus :: Lens' DescribePipelineResponse (Maybe PipelineStatus)

The status of the pipeline execution.

describePipelineResponse_pipelineDescription :: Lens' DescribePipelineResponse (Maybe Text)

The description of the pipeline.

describePipelineResponse_pipelineArn :: Lens' DescribePipelineResponse (Maybe Text)

The Amazon Resource Name (ARN) of the pipeline.

describePipelineResponse_pipelineDefinition :: Lens' DescribePipelineResponse (Maybe Text)

The JSON pipeline definition.

describePipelineResponse_lastModifiedBy :: Lens' DescribePipelineResponse (Maybe UserContext)

Undocumented member.

describePipelineResponse_roleArn :: Lens' DescribePipelineResponse (Maybe Text)

The Amazon Resource Name (ARN) that the pipeline uses to execute.

describePipelineResponse_httpStatus :: Lens' DescribePipelineResponse Int

The response's http status code.