Amazonka.SageMaker.ListDomains

Description

Lists the domains.

This operation returns paginated results.

Creating a Request

data ListDomains

See: newListDomains smart constructor.

Constructors

ListDomains'

Fields

nextToken :: Maybe Text

If the previous response was truncated, you will receive this token. Use it in your next request to receive the next set of results.

maxResults :: Maybe Natural

Returns a list up to a specified limit.

newListDomains :: ListDomains

Create a value of ListDomains with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDomains', listDomains_nextToken - If the previous response was truncated, you will receive this token. Use it in your next request to receive the next set of results.

$sel:maxResults:ListDomains', listDomains_maxResults - Returns a list up to a specified limit.

Request Lenses

listDomains_nextToken :: Lens' ListDomains (Maybe Text)

If the previous response was truncated, you will receive this token. Use it in your next request to receive the next set of results.

listDomains_maxResults :: Lens' ListDomains (Maybe Natural)

Returns a list up to a specified limit.

Destructuring the Response

data ListDomainsResponse

See: newListDomainsResponse smart constructor.

Constructors

ListDomainsResponse'

Fields

nextToken :: Maybe Text

If the previous response was truncated, you will receive this token. Use it in your next request to receive the next set of results.

domains :: Maybe [DomainDetails]

The list of domains.

httpStatus :: Int

The response's http status code.

newListDomainsResponse

Arguments

:: Int
$sel:httpStatus:ListDomainsResponse'

-> ListDomainsResponse

Create a value of ListDomainsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDomains', listDomainsResponse_nextToken - If the previous response was truncated, you will receive this token. Use it in your next request to receive the next set of results.

$sel:domains:ListDomainsResponse', listDomainsResponse_domains - The list of domains.

$sel:httpStatus:ListDomainsResponse', listDomainsResponse_httpStatus - The response's http status code.

Response Lenses

listDomainsResponse_nextToken :: Lens' ListDomainsResponse (Maybe Text)

If the previous response was truncated, you will receive this token. Use it in your next request to receive the next set of results.

listDomainsResponse_domains :: Lens' ListDomainsResponse (Maybe [DomainDetails])

The list of domains.

listDomainsResponse_httpStatus :: Lens' ListDomainsResponse Int

The response's http status code.