Amazonka.SageMaker.ListEdgePackagingJobs

Description

Returns a list of edge packaging jobs.

This operation returns paginated results.

Creating a Request

data ListEdgePackagingJobs

See: newListEdgePackagingJobs smart constructor.

Constructors

ListEdgePackagingJobs'

Fields

nameContains :: Maybe Text

Filter for jobs containing this name in their packaging job name.

lastModifiedTimeBefore :: Maybe POSIX

Select jobs where the job was updated before specified time.

creationTimeAfter :: Maybe POSIX

Select jobs where the job was created after specified time.

nextToken :: Maybe Text

The response from the last list when returning a list large enough to need tokening.

sortOrder :: Maybe SortOrder

What direction to sort by.

lastModifiedTimeAfter :: Maybe POSIX

Select jobs where the job was updated after specified time.

creationTimeBefore :: Maybe POSIX

Select jobs where the job was created before specified time.

modelNameContains :: Maybe Text

Filter for jobs where the model name contains this string.

statusEquals :: Maybe EdgePackagingJobStatus

The job status to filter for.

maxResults :: Maybe Int

Maximum number of results to select.

sortBy :: Maybe ListEdgePackagingJobsSortBy

Use to specify what column to sort by.

newListEdgePackagingJobs :: ListEdgePackagingJobs

Create a value of ListEdgePackagingJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nameContains:ListEdgePackagingJobs', listEdgePackagingJobs_nameContains - Filter for jobs containing this name in their packaging job name.

$sel:lastModifiedTimeBefore:ListEdgePackagingJobs', listEdgePackagingJobs_lastModifiedTimeBefore - Select jobs where the job was updated before specified time.

$sel:creationTimeAfter:ListEdgePackagingJobs', listEdgePackagingJobs_creationTimeAfter - Select jobs where the job was created after specified time.

$sel:nextToken:ListEdgePackagingJobs', listEdgePackagingJobs_nextToken - The response from the last list when returning a list large enough to need tokening.

$sel:sortOrder:ListEdgePackagingJobs', listEdgePackagingJobs_sortOrder - What direction to sort by.

$sel:lastModifiedTimeAfter:ListEdgePackagingJobs', listEdgePackagingJobs_lastModifiedTimeAfter - Select jobs where the job was updated after specified time.

$sel:creationTimeBefore:ListEdgePackagingJobs', listEdgePackagingJobs_creationTimeBefore - Select jobs where the job was created before specified time.

$sel:modelNameContains:ListEdgePackagingJobs', listEdgePackagingJobs_modelNameContains - Filter for jobs where the model name contains this string.

$sel:statusEquals:ListEdgePackagingJobs', listEdgePackagingJobs_statusEquals - The job status to filter for.

$sel:maxResults:ListEdgePackagingJobs', listEdgePackagingJobs_maxResults - Maximum number of results to select.

$sel:sortBy:ListEdgePackagingJobs', listEdgePackagingJobs_sortBy - Use to specify what column to sort by.

Request Lenses

listEdgePackagingJobs_nameContains :: Lens' ListEdgePackagingJobs (Maybe Text)

Filter for jobs containing this name in their packaging job name.

listEdgePackagingJobs_lastModifiedTimeBefore :: Lens' ListEdgePackagingJobs (Maybe UTCTime)

Select jobs where the job was updated before specified time.

listEdgePackagingJobs_creationTimeAfter :: Lens' ListEdgePackagingJobs (Maybe UTCTime)

Select jobs where the job was created after specified time.

listEdgePackagingJobs_nextToken :: Lens' ListEdgePackagingJobs (Maybe Text)

The response from the last list when returning a list large enough to need tokening.

listEdgePackagingJobs_sortOrder :: Lens' ListEdgePackagingJobs (Maybe SortOrder)

What direction to sort by.

listEdgePackagingJobs_lastModifiedTimeAfter :: Lens' ListEdgePackagingJobs (Maybe UTCTime)

Select jobs where the job was updated after specified time.

listEdgePackagingJobs_creationTimeBefore :: Lens' ListEdgePackagingJobs (Maybe UTCTime)

Select jobs where the job was created before specified time.

listEdgePackagingJobs_modelNameContains :: Lens' ListEdgePackagingJobs (Maybe Text)

Filter for jobs where the model name contains this string.

listEdgePackagingJobs_statusEquals :: Lens' ListEdgePackagingJobs (Maybe EdgePackagingJobStatus)

The job status to filter for.

listEdgePackagingJobs_maxResults :: Lens' ListEdgePackagingJobs (Maybe Int)

Maximum number of results to select.

listEdgePackagingJobs_sortBy :: Lens' ListEdgePackagingJobs (Maybe ListEdgePackagingJobsSortBy)

Use to specify what column to sort by.

Destructuring the Response

data ListEdgePackagingJobsResponse

See: newListEdgePackagingJobsResponse smart constructor.

Constructors

ListEdgePackagingJobsResponse'

Fields

nextToken :: Maybe Text

Token to use when calling the next page of results.

httpStatus :: Int

The response's http status code.

edgePackagingJobSummaries :: [EdgePackagingJobSummary]

Summaries of edge packaging jobs.

newListEdgePackagingJobsResponse

Arguments

:: Int
$sel:httpStatus:ListEdgePackagingJobsResponse'

-> ListEdgePackagingJobsResponse

Create a value of ListEdgePackagingJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEdgePackagingJobs', listEdgePackagingJobsResponse_nextToken - Token to use when calling the next page of results.

$sel:httpStatus:ListEdgePackagingJobsResponse', listEdgePackagingJobsResponse_httpStatus - The response's http status code.

$sel:edgePackagingJobSummaries:ListEdgePackagingJobsResponse', listEdgePackagingJobsResponse_edgePackagingJobSummaries - Summaries of edge packaging jobs.

Response Lenses

listEdgePackagingJobsResponse_nextToken :: Lens' ListEdgePackagingJobsResponse (Maybe Text)

Token to use when calling the next page of results.

listEdgePackagingJobsResponse_httpStatus :: Lens' ListEdgePackagingJobsResponse Int

The response's http status code.

listEdgePackagingJobsResponse_edgePackagingJobSummaries :: Lens' ListEdgePackagingJobsResponse [EdgePackagingJobSummary]

Summaries of edge packaging jobs.