Amazonka.SageMaker.ListModelBiasJobDefinitions

Description

Lists model bias jobs definitions that satisfy various filters.

This operation returns paginated results.

Creating a Request

data ListModelBiasJobDefinitions

See: newListModelBiasJobDefinitions smart constructor.

Constructors

ListModelBiasJobDefinitions'

Fields

nameContains :: Maybe Text

Filter for model bias jobs whose name contains a specified string.

endpointName :: Maybe Text

Name of the endpoint to monitor for model bias.

creationTimeAfter :: Maybe POSIX

A filter that returns only model bias jobs created after a specified time.

nextToken :: Maybe Text

The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

sortOrder :: Maybe SortOrder

Whether to sort the results in Ascending or Descending order. The default is Descending.

creationTimeBefore :: Maybe POSIX

A filter that returns only model bias jobs created before a specified time.

maxResults :: Maybe Natural

The maximum number of model bias jobs to return in the response. The default value is 10.

sortBy :: Maybe MonitoringJobDefinitionSortKey

Whether to sort results by the Name or CreationTime field. The default is CreationTime.

newListModelBiasJobDefinitions :: ListModelBiasJobDefinitions

Create a value of ListModelBiasJobDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nameContains:ListModelBiasJobDefinitions', listModelBiasJobDefinitions_nameContains - Filter for model bias jobs whose name contains a specified string.

$sel:endpointName:ListModelBiasJobDefinitions', listModelBiasJobDefinitions_endpointName - Name of the endpoint to monitor for model bias.

$sel:creationTimeAfter:ListModelBiasJobDefinitions', listModelBiasJobDefinitions_creationTimeAfter - A filter that returns only model bias jobs created after a specified time.

$sel:nextToken:ListModelBiasJobDefinitions', listModelBiasJobDefinitions_nextToken - The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

$sel:sortOrder:ListModelBiasJobDefinitions', listModelBiasJobDefinitions_sortOrder - Whether to sort the results in Ascending or Descending order. The default is Descending.

$sel:creationTimeBefore:ListModelBiasJobDefinitions', listModelBiasJobDefinitions_creationTimeBefore - A filter that returns only model bias jobs created before a specified time.

$sel:maxResults:ListModelBiasJobDefinitions', listModelBiasJobDefinitions_maxResults - The maximum number of model bias jobs to return in the response. The default value is 10.

$sel:sortBy:ListModelBiasJobDefinitions', listModelBiasJobDefinitions_sortBy - Whether to sort results by the Name or CreationTime field. The default is CreationTime.

Request Lenses

listModelBiasJobDefinitions_nameContains :: Lens' ListModelBiasJobDefinitions (Maybe Text)

Filter for model bias jobs whose name contains a specified string.

listModelBiasJobDefinitions_endpointName :: Lens' ListModelBiasJobDefinitions (Maybe Text)

Name of the endpoint to monitor for model bias.

listModelBiasJobDefinitions_creationTimeAfter :: Lens' ListModelBiasJobDefinitions (Maybe UTCTime)

A filter that returns only model bias jobs created after a specified time.

listModelBiasJobDefinitions_nextToken :: Lens' ListModelBiasJobDefinitions (Maybe Text)

The token returned if the response is truncated. To retrieve the next set of job executions, use it in the next request.

listModelBiasJobDefinitions_sortOrder :: Lens' ListModelBiasJobDefinitions (Maybe SortOrder)

Whether to sort the results in Ascending or Descending order. The default is Descending.

listModelBiasJobDefinitions_creationTimeBefore :: Lens' ListModelBiasJobDefinitions (Maybe UTCTime)

A filter that returns only model bias jobs created before a specified time.

listModelBiasJobDefinitions_maxResults :: Lens' ListModelBiasJobDefinitions (Maybe Natural)

The maximum number of model bias jobs to return in the response. The default value is 10.

listModelBiasJobDefinitions_sortBy :: Lens' ListModelBiasJobDefinitions (Maybe MonitoringJobDefinitionSortKey)

Whether to sort results by the Name or CreationTime field. The default is CreationTime.

Destructuring the Response

data ListModelBiasJobDefinitionsResponse

See: newListModelBiasJobDefinitionsResponse smart constructor.

Constructors

ListModelBiasJobDefinitionsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of jobs, use it in the subsequent request.

httpStatus :: Int

The response's http status code.

jobDefinitionSummaries :: [MonitoringJobDefinitionSummary]

A JSON array in which each element is a summary for a model bias jobs.

newListModelBiasJobDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:ListModelBiasJobDefinitionsResponse'

-> ListModelBiasJobDefinitionsResponse

Create a value of ListModelBiasJobDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListModelBiasJobDefinitions', listModelBiasJobDefinitionsResponse_nextToken - If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of jobs, use it in the subsequent request.

$sel:httpStatus:ListModelBiasJobDefinitionsResponse', listModelBiasJobDefinitionsResponse_httpStatus - The response's http status code.

$sel:jobDefinitionSummaries:ListModelBiasJobDefinitionsResponse', listModelBiasJobDefinitionsResponse_jobDefinitionSummaries - A JSON array in which each element is a summary for a model bias jobs.

Response Lenses

listModelBiasJobDefinitionsResponse_nextToken :: Lens' ListModelBiasJobDefinitionsResponse (Maybe Text)

If the response is truncated, Amazon SageMaker returns this token. To retrieve the next set of jobs, use it in the subsequent request.

listModelBiasJobDefinitionsResponse_httpStatus :: Lens' ListModelBiasJobDefinitionsResponse Int

The response's http status code.

listModelBiasJobDefinitionsResponse_jobDefinitionSummaries :: Lens' ListModelBiasJobDefinitionsResponse [MonitoringJobDefinitionSummary]

A JSON array in which each element is a summary for a model bias jobs.