Amazonka.SageMaker.ListPipelines

Description

Gets a list of pipelines.

This operation returns paginated results.

Creating a Request

data ListPipelines

See: newListPipelines smart constructor.

Constructors

ListPipelines'

Fields

createdAfter :: Maybe POSIX

A filter that returns the pipelines that were created after a specified time.

pipelineNamePrefix :: Maybe Text

The prefix of the pipeline name.

nextToken :: Maybe Text

If the result of the previous ListPipelines request was truncated, the response includes a NextToken. To retrieve the next set of pipelines, use the token in the next request.

sortOrder :: Maybe SortOrder

The sort order for results.

maxResults :: Maybe Natural

The maximum number of pipelines to return in the response.

createdBefore :: Maybe POSIX

A filter that returns the pipelines that were created before a specified time.

sortBy :: Maybe SortPipelinesBy

The field by which to sort results. The default is CreatedTime.

newListPipelines :: ListPipelines

Create a value of ListPipelines with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAfter:ListPipelines', listPipelines_createdAfter - A filter that returns the pipelines that were created after a specified time.

$sel:pipelineNamePrefix:ListPipelines', listPipelines_pipelineNamePrefix - The prefix of the pipeline name.

$sel:nextToken:ListPipelines', listPipelines_nextToken - If the result of the previous ListPipelines request was truncated, the response includes a NextToken. To retrieve the next set of pipelines, use the token in the next request.

$sel:sortOrder:ListPipelines', listPipelines_sortOrder - The sort order for results.

$sel:maxResults:ListPipelines', listPipelines_maxResults - The maximum number of pipelines to return in the response.

$sel:createdBefore:ListPipelines', listPipelines_createdBefore - A filter that returns the pipelines that were created before a specified time.

$sel:sortBy:ListPipelines', listPipelines_sortBy - The field by which to sort results. The default is CreatedTime.

Request Lenses

listPipelines_createdAfter :: Lens' ListPipelines (Maybe UTCTime)

A filter that returns the pipelines that were created after a specified time.

listPipelines_pipelineNamePrefix :: Lens' ListPipelines (Maybe Text)

The prefix of the pipeline name.

listPipelines_nextToken :: Lens' ListPipelines (Maybe Text)

If the result of the previous ListPipelines request was truncated, the response includes a NextToken. To retrieve the next set of pipelines, use the token in the next request.

listPipelines_sortOrder :: Lens' ListPipelines (Maybe SortOrder)

The sort order for results.

listPipelines_maxResults :: Lens' ListPipelines (Maybe Natural)

The maximum number of pipelines to return in the response.

listPipelines_createdBefore :: Lens' ListPipelines (Maybe UTCTime)

A filter that returns the pipelines that were created before a specified time.

listPipelines_sortBy :: Lens' ListPipelines (Maybe SortPipelinesBy)

The field by which to sort results. The default is CreatedTime.

Destructuring the Response

data ListPipelinesResponse

See: newListPipelinesResponse smart constructor.

Constructors

ListPipelinesResponse'

Fields

pipelineSummaries :: Maybe [PipelineSummary]

Contains a sorted list of PipelineSummary objects matching the specified filters. Each PipelineSummary consists of PipelineArn, PipelineName, ExperimentName, PipelineDescription, CreationTime, LastModifiedTime, LastRunTime, and RoleArn. This list can be empty.

nextToken :: Maybe Text

If the result of the previous ListPipelines request was truncated, the response includes a NextToken. To retrieve the next set of pipelines, use the token in the next request.

httpStatus :: Int

The response's http status code.

newListPipelinesResponse

Arguments

:: Int
$sel:httpStatus:ListPipelinesResponse'

-> ListPipelinesResponse

Create a value of ListPipelinesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineSummaries:ListPipelinesResponse', listPipelinesResponse_pipelineSummaries - Contains a sorted list of PipelineSummary objects matching the specified filters. Each PipelineSummary consists of PipelineArn, PipelineName, ExperimentName, PipelineDescription, CreationTime, LastModifiedTime, LastRunTime, and RoleArn. This list can be empty.

$sel:nextToken:ListPipelines', listPipelinesResponse_nextToken - If the result of the previous ListPipelines request was truncated, the response includes a NextToken. To retrieve the next set of pipelines, use the token in the next request.

$sel:httpStatus:ListPipelinesResponse', listPipelinesResponse_httpStatus - The response's http status code.

Response Lenses

listPipelinesResponse_pipelineSummaries :: Lens' ListPipelinesResponse (Maybe [PipelineSummary])

Contains a sorted list of PipelineSummary objects matching the specified filters. Each PipelineSummary consists of PipelineArn, PipelineName, ExperimentName, PipelineDescription, CreationTime, LastModifiedTime, LastRunTime, and RoleArn. This list can be empty.

listPipelinesResponse_nextToken :: Lens' ListPipelinesResponse (Maybe Text)

If the result of the previous ListPipelines request was truncated, the response includes a NextToken. To retrieve the next set of pipelines, use the token in the next request.

listPipelinesResponse_httpStatus :: Lens' ListPipelinesResponse Int

The response's http status code.