Amazonka.SageMaker.ListTrials

Description

Lists the trials in your account. Specify an experiment name to limit the list to the trials that are part of that experiment. Specify a trial component name to limit the list to the trials that associated with that trial component. The list can be filtered to show only trials that were created in a specific time range. The list can be sorted by trial name or creation time.

This operation returns paginated results.

Creating a Request

data ListTrials

See: newListTrials smart constructor.

Constructors

ListTrials'

Fields

createdAfter :: Maybe POSIX

A filter that returns only trials created after the specified time.

experimentName :: Maybe Text

A filter that returns only trials that are part of the specified experiment.

nextToken :: Maybe Text

If the previous call to ListTrials didn't return the full set of trials, the call returns a token for getting the next set of trials.

sortOrder :: Maybe SortOrder

The sort order. The default value is Descending.

trialComponentName :: Maybe Text

A filter that returns only trials that are associated with the specified trial component.

maxResults :: Maybe Natural

The maximum number of trials to return in the response. The default value is 10.

createdBefore :: Maybe POSIX

A filter that returns only trials created before the specified time.

sortBy :: Maybe SortTrialsBy

The property used to sort results. The default value is CreationTime.

newListTrials :: ListTrials

Create a value of ListTrials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAfter:ListTrials', listTrials_createdAfter - A filter that returns only trials created after the specified time.

$sel:experimentName:ListTrials', listTrials_experimentName - A filter that returns only trials that are part of the specified experiment.

$sel:nextToken:ListTrials', listTrials_nextToken - If the previous call to ListTrials didn't return the full set of trials, the call returns a token for getting the next set of trials.

$sel:sortOrder:ListTrials', listTrials_sortOrder - The sort order. The default value is Descending.

$sel:trialComponentName:ListTrials', listTrials_trialComponentName - A filter that returns only trials that are associated with the specified trial component.

$sel:maxResults:ListTrials', listTrials_maxResults - The maximum number of trials to return in the response. The default value is 10.

$sel:createdBefore:ListTrials', listTrials_createdBefore - A filter that returns only trials created before the specified time.

$sel:sortBy:ListTrials', listTrials_sortBy - The property used to sort results. The default value is CreationTime.

Request Lenses

listTrials_createdAfter :: Lens' ListTrials (Maybe UTCTime)

A filter that returns only trials created after the specified time.

listTrials_experimentName :: Lens' ListTrials (Maybe Text)

A filter that returns only trials that are part of the specified experiment.

listTrials_nextToken :: Lens' ListTrials (Maybe Text)

If the previous call to ListTrials didn't return the full set of trials, the call returns a token for getting the next set of trials.

listTrials_sortOrder :: Lens' ListTrials (Maybe SortOrder)

The sort order. The default value is Descending.

listTrials_trialComponentName :: Lens' ListTrials (Maybe Text)

A filter that returns only trials that are associated with the specified trial component.

listTrials_maxResults :: Lens' ListTrials (Maybe Natural)

The maximum number of trials to return in the response. The default value is 10.

listTrials_createdBefore :: Lens' ListTrials (Maybe UTCTime)

A filter that returns only trials created before the specified time.

listTrials_sortBy :: Lens' ListTrials (Maybe SortTrialsBy)

The property used to sort results. The default value is CreationTime.

Destructuring the Response

data ListTrialsResponse

See: newListTrialsResponse smart constructor.

Constructors

ListTrialsResponse'

Fields

nextToken :: Maybe Text

A token for getting the next set of trials, if there are any.

trialSummaries :: Maybe [TrialSummary]

A list of the summaries of your trials.

httpStatus :: Int

The response's http status code.

newListTrialsResponse

Arguments

:: Int
$sel:httpStatus:ListTrialsResponse'

-> ListTrialsResponse

Create a value of ListTrialsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTrials', listTrialsResponse_nextToken - A token for getting the next set of trials, if there are any.

$sel:trialSummaries:ListTrialsResponse', listTrialsResponse_trialSummaries - A list of the summaries of your trials.

$sel:httpStatus:ListTrialsResponse', listTrialsResponse_httpStatus - The response's http status code.

Response Lenses

listTrialsResponse_nextToken :: Lens' ListTrialsResponse (Maybe Text)

A token for getting the next set of trials, if there are any.

listTrialsResponse_trialSummaries :: Lens' ListTrialsResponse (Maybe [TrialSummary])

A list of the summaries of your trials.

listTrialsResponse_httpStatus :: Lens' ListTrialsResponse Int

The response's http status code.