Amazonka.SageMaker.Types.CodeRepositorySummary

Description

Documentation

data CodeRepositorySummary

Specifies summary information about a Git repository.

See: newCodeRepositorySummary smart constructor.

Constructors

CodeRepositorySummary'

Fields

gitConfig :: Maybe GitConfig

Configuration details for the Git repository, including the URL where it is located and the ARN of the Amazon Web Services Secrets Manager secret that contains the credentials used to access the repository.

codeRepositoryName :: Text

The name of the Git repository.

codeRepositoryArn :: Text

The Amazon Resource Name (ARN) of the Git repository.

creationTime :: POSIX

The date and time that the Git repository was created.

lastModifiedTime :: POSIX

The date and time that the Git repository was last modified.

newCodeRepositorySummary

Arguments

:: Text
$sel:codeRepositoryName:CodeRepositorySummary'

-> Text
$sel:codeRepositoryArn:CodeRepositorySummary'

-> UTCTime
$sel:creationTime:CodeRepositorySummary'

-> UTCTime
$sel:lastModifiedTime:CodeRepositorySummary'

-> CodeRepositorySummary

Create a value of CodeRepositorySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gitConfig:CodeRepositorySummary', codeRepositorySummary_gitConfig - Configuration details for the Git repository, including the URL where it is located and the ARN of the Amazon Web Services Secrets Manager secret that contains the credentials used to access the repository.

$sel:codeRepositoryName:CodeRepositorySummary', codeRepositorySummary_codeRepositoryName - The name of the Git repository.

$sel:codeRepositoryArn:CodeRepositorySummary', codeRepositorySummary_codeRepositoryArn - The Amazon Resource Name (ARN) of the Git repository.

$sel:creationTime:CodeRepositorySummary', codeRepositorySummary_creationTime - The date and time that the Git repository was created.

$sel:lastModifiedTime:CodeRepositorySummary', codeRepositorySummary_lastModifiedTime - The date and time that the Git repository was last modified.

codeRepositorySummary_gitConfig :: Lens' CodeRepositorySummary (Maybe GitConfig)

Configuration details for the Git repository, including the URL where it is located and the ARN of the Amazon Web Services Secrets Manager secret that contains the credentials used to access the repository.

codeRepositorySummary_codeRepositoryName :: Lens' CodeRepositorySummary Text

The name of the Git repository.

codeRepositorySummary_codeRepositoryArn :: Lens' CodeRepositorySummary Text

The Amazon Resource Name (ARN) of the Git repository.

codeRepositorySummary_creationTime :: Lens' CodeRepositorySummary UTCTime

The date and time that the Git repository was created.

codeRepositorySummary_lastModifiedTime :: Lens' CodeRepositorySummary UTCTime

The date and time that the Git repository was last modified.