Amazonka.SageMaker.Types.DataQualityAppSpecification

Description

Documentation

data DataQualityAppSpecification

Information about the container that a data quality monitoring job runs.

See: newDataQualityAppSpecification smart constructor.

Constructors

DataQualityAppSpecification'

Fields

containerArguments :: Maybe (NonEmpty Text)

The arguments to send to the container that the monitoring job runs.

recordPreprocessorSourceUri :: Maybe Text

An Amazon S3 URI to a script that is called per row prior to running analysis. It can base64 decode the payload and convert it into a flatted json so that the built-in container can use the converted data. Applicable only for the built-in (first party) containers.

environment :: Maybe (HashMap Text Text)

Sets the environment variables in the container that the monitoring job runs.

containerEntrypoint :: Maybe (NonEmpty Text)

The entrypoint for a container used to run a monitoring job.

postAnalyticsProcessorSourceUri :: Maybe Text

An Amazon S3 URI to a script that is called after analysis has been performed. Applicable only for the built-in (first party) containers.

imageUri :: Text

The container image that the data quality monitoring job runs.

newDataQualityAppSpecification

Arguments

:: Text
$sel:imageUri:DataQualityAppSpecification'

-> DataQualityAppSpecification

Create a value of DataQualityAppSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerArguments:DataQualityAppSpecification', dataQualityAppSpecification_containerArguments - The arguments to send to the container that the monitoring job runs.

$sel:recordPreprocessorSourceUri:DataQualityAppSpecification', dataQualityAppSpecification_recordPreprocessorSourceUri - An Amazon S3 URI to a script that is called per row prior to running analysis. It can base64 decode the payload and convert it into a flatted json so that the built-in container can use the converted data. Applicable only for the built-in (first party) containers.

$sel:environment:DataQualityAppSpecification', dataQualityAppSpecification_environment - Sets the environment variables in the container that the monitoring job runs.

$sel:containerEntrypoint:DataQualityAppSpecification', dataQualityAppSpecification_containerEntrypoint - The entrypoint for a container used to run a monitoring job.

$sel:postAnalyticsProcessorSourceUri:DataQualityAppSpecification', dataQualityAppSpecification_postAnalyticsProcessorSourceUri - An Amazon S3 URI to a script that is called after analysis has been performed. Applicable only for the built-in (first party) containers.

$sel:imageUri:DataQualityAppSpecification', dataQualityAppSpecification_imageUri - The container image that the data quality monitoring job runs.

dataQualityAppSpecification_containerArguments :: Lens' DataQualityAppSpecification (Maybe (NonEmpty Text))

The arguments to send to the container that the monitoring job runs.

dataQualityAppSpecification_recordPreprocessorSourceUri :: Lens' DataQualityAppSpecification (Maybe Text)

An Amazon S3 URI to a script that is called per row prior to running analysis. It can base64 decode the payload and convert it into a flatted json so that the built-in container can use the converted data. Applicable only for the built-in (first party) containers.

dataQualityAppSpecification_environment :: Lens' DataQualityAppSpecification (Maybe (HashMap Text Text))

Sets the environment variables in the container that the monitoring job runs.

dataQualityAppSpecification_containerEntrypoint :: Lens' DataQualityAppSpecification (Maybe (NonEmpty Text))

The entrypoint for a container used to run a monitoring job.

dataQualityAppSpecification_postAnalyticsProcessorSourceUri :: Lens' DataQualityAppSpecification (Maybe Text)

An Amazon S3 URI to a script that is called after analysis has been performed. Applicable only for the built-in (first party) containers.

dataQualityAppSpecification_imageUri :: Lens' DataQualityAppSpecification Text

The container image that the data quality monitoring job runs.