Amazonka.SageMaker.Types.DeviceSummary

Description

Documentation

data DeviceSummary

Summary of the device.

See: newDeviceSummary smart constructor.

Constructors

DeviceSummary'

Fields

registrationTime :: Maybe POSIX

The timestamp of the last registration or de-reregistration.

models :: Maybe [EdgeModelSummary]

Models on the device.

latestHeartbeat :: Maybe POSIX

The last heartbeat received from the device.

description :: Maybe Text

A description of the device.

deviceFleetName :: Maybe Text

The name of the fleet the device belongs to.

iotThingName :: Maybe Text

The Amazon Web Services Internet of Things (IoT) object thing name associated with the device..

deviceName :: Text

The unique identifier of the device.

deviceArn :: Text

Amazon Resource Name (ARN) of the device.

newDeviceSummary

Arguments

:: Text
$sel:deviceName:DeviceSummary'

-> Text
$sel:deviceArn:DeviceSummary'

-> DeviceSummary

Create a value of DeviceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registrationTime:DeviceSummary', deviceSummary_registrationTime - The timestamp of the last registration or de-reregistration.

$sel:models:DeviceSummary', deviceSummary_models - Models on the device.

$sel:latestHeartbeat:DeviceSummary', deviceSummary_latestHeartbeat - The last heartbeat received from the device.

$sel:description:DeviceSummary', deviceSummary_description - A description of the device.

$sel:deviceFleetName:DeviceSummary', deviceSummary_deviceFleetName - The name of the fleet the device belongs to.

$sel:iotThingName:DeviceSummary', deviceSummary_iotThingName - The Amazon Web Services Internet of Things (IoT) object thing name associated with the device..

$sel:deviceName:DeviceSummary', deviceSummary_deviceName - The unique identifier of the device.

$sel:deviceArn:DeviceSummary', deviceSummary_deviceArn - Amazon Resource Name (ARN) of the device.

deviceSummary_registrationTime :: Lens' DeviceSummary (Maybe UTCTime)

The timestamp of the last registration or de-reregistration.

deviceSummary_models :: Lens' DeviceSummary (Maybe [EdgeModelSummary])

Models on the device.

deviceSummary_latestHeartbeat :: Lens' DeviceSummary (Maybe UTCTime)

The last heartbeat received from the device.

deviceSummary_description :: Lens' DeviceSummary (Maybe Text)

A description of the device.

deviceSummary_deviceFleetName :: Lens' DeviceSummary (Maybe Text)

The name of the fleet the device belongs to.

deviceSummary_iotThingName :: Lens' DeviceSummary (Maybe Text)

The Amazon Web Services Internet of Things (IoT) object thing name associated with the device..

deviceSummary_deviceName :: Lens' DeviceSummary Text

The unique identifier of the device.

deviceSummary_deviceArn :: Lens' DeviceSummary Text

Amazon Resource Name (ARN) of the device.