Amazonka.SageMaker.Types.EdgePackagingJobSummary

Description

Documentation

data EdgePackagingJobSummary

Summary of edge packaging job.

See: newEdgePackagingJobSummary smart constructor.

Constructors

EdgePackagingJobSummary'

Fields

creationTime :: Maybe POSIX

The timestamp of when the job was created.

modelName :: Maybe Text

The name of the model.

lastModifiedTime :: Maybe POSIX

The timestamp of when the edge packaging job was last updated.

compilationJobName :: Maybe Text

The name of the SageMaker Neo compilation job.

modelVersion :: Maybe Text

The version of the model.

edgePackagingJobArn :: Text

The Amazon Resource Name (ARN) of the edge packaging job.

edgePackagingJobName :: Text

The name of the edge packaging job.

edgePackagingJobStatus :: EdgePackagingJobStatus

The status of the edge packaging job.

newEdgePackagingJobSummary

Arguments

:: Text
$sel:edgePackagingJobArn:EdgePackagingJobSummary'

-> Text
$sel:edgePackagingJobName:EdgePackagingJobSummary'

-> EdgePackagingJobStatus
$sel:edgePackagingJobStatus:EdgePackagingJobSummary'

-> EdgePackagingJobSummary

Create a value of EdgePackagingJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:EdgePackagingJobSummary', edgePackagingJobSummary_creationTime - The timestamp of when the job was created.

$sel:modelName:EdgePackagingJobSummary', edgePackagingJobSummary_modelName - The name of the model.

$sel:lastModifiedTime:EdgePackagingJobSummary', edgePackagingJobSummary_lastModifiedTime - The timestamp of when the edge packaging job was last updated.

$sel:compilationJobName:EdgePackagingJobSummary', edgePackagingJobSummary_compilationJobName - The name of the SageMaker Neo compilation job.

$sel:modelVersion:EdgePackagingJobSummary', edgePackagingJobSummary_modelVersion - The version of the model.

$sel:edgePackagingJobArn:EdgePackagingJobSummary', edgePackagingJobSummary_edgePackagingJobArn - The Amazon Resource Name (ARN) of the edge packaging job.

$sel:edgePackagingJobName:EdgePackagingJobSummary', edgePackagingJobSummary_edgePackagingJobName - The name of the edge packaging job.

$sel:edgePackagingJobStatus:EdgePackagingJobSummary', edgePackagingJobSummary_edgePackagingJobStatus - The status of the edge packaging job.

edgePackagingJobSummary_creationTime :: Lens' EdgePackagingJobSummary (Maybe UTCTime)

The timestamp of when the job was created.

edgePackagingJobSummary_modelName :: Lens' EdgePackagingJobSummary (Maybe Text)

The name of the model.

edgePackagingJobSummary_lastModifiedTime :: Lens' EdgePackagingJobSummary (Maybe UTCTime)

The timestamp of when the edge packaging job was last updated.

edgePackagingJobSummary_compilationJobName :: Lens' EdgePackagingJobSummary (Maybe Text)

The name of the SageMaker Neo compilation job.

edgePackagingJobSummary_modelVersion :: Lens' EdgePackagingJobSummary (Maybe Text)

The version of the model.

edgePackagingJobSummary_edgePackagingJobArn :: Lens' EdgePackagingJobSummary Text

The Amazon Resource Name (ARN) of the edge packaging job.

edgePackagingJobSummary_edgePackagingJobName :: Lens' EdgePackagingJobSummary Text

The name of the edge packaging job.

edgePackagingJobSummary_edgePackagingJobStatus :: Lens' EdgePackagingJobSummary EdgePackagingJobStatus

The status of the edge packaging job.