Amazonka.SageMaker.Types.ExperimentConfig

Description

Documentation

data ExperimentConfig

Associates a SageMaker job as a trial component with an experiment and trial. Specified when you call the following APIs:

CreateProcessingJob

CreateTrainingJob

CreateTransformJob

See: newExperimentConfig smart constructor.

Constructors

ExperimentConfig'

Fields

trialComponentDisplayName :: Maybe Text

The display name for the trial component. If this key isn't specified, the display name is the trial component name.

experimentName :: Maybe Text

The name of an existing experiment to associate the trial component with.

trialName :: Maybe Text

The name of an existing trial to associate the trial component with. If not specified, a new trial is created.

newExperimentConfig :: ExperimentConfig

Create a value of ExperimentConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trialComponentDisplayName:ExperimentConfig', experimentConfig_trialComponentDisplayName - The display name for the trial component. If this key isn't specified, the display name is the trial component name.

$sel:experimentName:ExperimentConfig', experimentConfig_experimentName - The name of an existing experiment to associate the trial component with.

$sel:trialName:ExperimentConfig', experimentConfig_trialName - The name of an existing trial to associate the trial component with. If not specified, a new trial is created.

experimentConfig_trialComponentDisplayName :: Lens' ExperimentConfig (Maybe Text)

The display name for the trial component. If this key isn't specified, the display name is the trial component name.

experimentConfig_experimentName :: Lens' ExperimentConfig (Maybe Text)

The name of an existing experiment to associate the trial component with.

experimentConfig_trialName :: Lens' ExperimentConfig (Maybe Text)

The name of an existing trial to associate the trial component with. If not specified, a new trial is created.