Amazonka.SageMaker.Types.FeatureGroup

Description

Documentation

data FeatureGroup

Amazon SageMaker Feature Store stores features in a collection called Feature Group. A Feature Group can be visualized as a table which has rows, with a unique identifier for each row where each column in the table is a feature. In principle, a Feature Group is composed of features and values per features.

See: newFeatureGroup smart constructor.

Constructors

FeatureGroup'

Fields

creationTime :: Maybe POSIX

The time a FeatureGroup was created.

offlineStoreConfig :: Maybe OfflineStoreConfig

failureReason :: Maybe Text

The reason that the FeatureGroup failed to be replicated in the OfflineStore. This is failure may be due to a failure to create a FeatureGroup in or delete a FeatureGroup from the OfflineStore.

featureGroupStatus :: Maybe FeatureGroupStatus

A FeatureGroup status.

featureDefinitions :: Maybe (NonEmpty FeatureDefinition)

A list of Features. Each Feature must include a FeatureName and a FeatureType.

Valid FeatureTypes are Integral, Fractional and String.

FeatureNames cannot be any of the following: is_deleted, write_time, api_invocation_time.

You can create up to 2,500 FeatureDefinitions per FeatureGroup.

offlineStoreStatus :: Maybe OfflineStoreStatus

onlineStoreConfig :: Maybe OnlineStoreConfig

eventTimeFeatureName :: Maybe Text

The name of the feature that stores the EventTime of a Record in a FeatureGroup.

A EventTime is point in time when a new event occurs that corresponds to the creation or update of a Record in FeatureGroup. All Records in the FeatureGroup must have a corresponding EventTime.

recordIdentifierFeatureName :: Maybe Text

The name of the Feature whose value uniquely identifies a Record defined in the FeatureGroup FeatureDefinitions.

featureGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of a FeatureGroup.

featureGroupName :: Maybe Text

The name of the FeatureGroup.

description :: Maybe Text

A free form description of a FeatureGroup.

tags :: Maybe [Tag]

Tags used to define a FeatureGroup.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM execution role used to create the feature group.

newFeatureGroup :: FeatureGroup

Create a value of FeatureGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:FeatureGroup', featureGroup_creationTime - The time a FeatureGroup was created.

$sel:offlineStoreConfig:FeatureGroup', featureGroup_offlineStoreConfig - Undocumented member.

$sel:failureReason:FeatureGroup', featureGroup_failureReason - The reason that the FeatureGroup failed to be replicated in the OfflineStore. This is failure may be due to a failure to create a FeatureGroup in or delete a FeatureGroup from the OfflineStore.

$sel:featureGroupStatus:FeatureGroup', featureGroup_featureGroupStatus - A FeatureGroup status.

$sel:featureDefinitions:FeatureGroup', featureGroup_featureDefinitions - A list of Features. Each Feature must include a FeatureName and a FeatureType.

Valid FeatureTypes are Integral, Fractional and String.

FeatureNames cannot be any of the following: is_deleted, write_time, api_invocation_time.

You can create up to 2,500 FeatureDefinitions per FeatureGroup.

$sel:offlineStoreStatus:FeatureGroup', featureGroup_offlineStoreStatus - Undocumented member.

$sel:onlineStoreConfig:FeatureGroup', featureGroup_onlineStoreConfig - Undocumented member.

$sel:eventTimeFeatureName:FeatureGroup', featureGroup_eventTimeFeatureName - The name of the feature that stores the EventTime of a Record in a FeatureGroup.

A EventTime is point in time when a new event occurs that corresponds to the creation or update of a Record in FeatureGroup. All Records in the FeatureGroup must have a corresponding EventTime.

$sel:recordIdentifierFeatureName:FeatureGroup', featureGroup_recordIdentifierFeatureName - The name of the Feature whose value uniquely identifies a Record defined in the FeatureGroup FeatureDefinitions.

$sel:featureGroupArn:FeatureGroup', featureGroup_featureGroupArn - The Amazon Resource Name (ARN) of a FeatureGroup.

$sel:featureGroupName:FeatureGroup', featureGroup_featureGroupName - The name of the FeatureGroup.

$sel:description:FeatureGroup', featureGroup_description - A free form description of a FeatureGroup.

$sel:tags:FeatureGroup', featureGroup_tags - Tags used to define a FeatureGroup.

$sel:roleArn:FeatureGroup', featureGroup_roleArn - The Amazon Resource Name (ARN) of the IAM execution role used to create the feature group.

featureGroup_creationTime :: Lens' FeatureGroup (Maybe UTCTime)

The time a FeatureGroup was created.

featureGroup_offlineStoreConfig :: Lens' FeatureGroup (Maybe OfflineStoreConfig)

Undocumented member.

featureGroup_failureReason :: Lens' FeatureGroup (Maybe Text)

The reason that the FeatureGroup failed to be replicated in the OfflineStore. This is failure may be due to a failure to create a FeatureGroup in or delete a FeatureGroup from the OfflineStore.

featureGroup_featureGroupStatus :: Lens' FeatureGroup (Maybe FeatureGroupStatus)

A FeatureGroup status.

featureGroup_featureDefinitions :: Lens' FeatureGroup (Maybe (NonEmpty FeatureDefinition))

A list of Features. Each Feature must include a FeatureName and a FeatureType.

Valid FeatureTypes are Integral, Fractional and String.

FeatureNames cannot be any of the following: is_deleted, write_time, api_invocation_time.

You can create up to 2,500 FeatureDefinitions per FeatureGroup.

featureGroup_offlineStoreStatus :: Lens' FeatureGroup (Maybe OfflineStoreStatus)

Undocumented member.

featureGroup_onlineStoreConfig :: Lens' FeatureGroup (Maybe OnlineStoreConfig)

Undocumented member.

featureGroup_eventTimeFeatureName :: Lens' FeatureGroup (Maybe Text)

The name of the feature that stores the EventTime of a Record in a FeatureGroup.

A EventTime is point in time when a new event occurs that corresponds to the creation or update of a Record in FeatureGroup. All Records in the FeatureGroup must have a corresponding EventTime.

featureGroup_recordIdentifierFeatureName :: Lens' FeatureGroup (Maybe Text)

The name of the Feature whose value uniquely identifies a Record defined in the FeatureGroup FeatureDefinitions.

featureGroup_featureGroupArn :: Lens' FeatureGroup (Maybe Text)

The Amazon Resource Name (ARN) of a FeatureGroup.

featureGroup_featureGroupName :: Lens' FeatureGroup (Maybe Text)

The name of the FeatureGroup.

featureGroup_description :: Lens' FeatureGroup (Maybe Text)

A free form description of a FeatureGroup.

featureGroup_tags :: Lens' FeatureGroup (Maybe [Tag])

Tags used to define a FeatureGroup.

featureGroup_roleArn :: Lens' FeatureGroup (Maybe Text)

The Amazon Resource Name (ARN) of the IAM execution role used to create the feature group.