Amazonka.SageMaker.Types.FileSystemDataSource

Description

Documentation

data FileSystemDataSource

Specifies a file system data source for a channel.

See: newFileSystemDataSource smart constructor.

Constructors

FileSystemDataSource'

Fields

fileSystemId :: Text

The file system id.

fileSystemAccessMode :: FileSystemAccessMode

The access mode of the mount of the directory associated with the channel. A directory can be mounted either in ro (read-only) or rw (read-write) mode.

fileSystemType :: FileSystemType

The file system type.

directoryPath :: Text

The full path to the directory to associate with the channel.

newFileSystemDataSource

Arguments

:: Text
$sel:fileSystemId:FileSystemDataSource'

-> FileSystemAccessMode
$sel:fileSystemAccessMode:FileSystemDataSource'

-> FileSystemType
$sel:fileSystemType:FileSystemDataSource'

-> Text
$sel:directoryPath:FileSystemDataSource'

-> FileSystemDataSource

Create a value of FileSystemDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileSystemId:FileSystemDataSource', fileSystemDataSource_fileSystemId - The file system id.

$sel:fileSystemAccessMode:FileSystemDataSource', fileSystemDataSource_fileSystemAccessMode - The access mode of the mount of the directory associated with the channel. A directory can be mounted either in ro (read-only) or rw (read-write) mode.

$sel:fileSystemType:FileSystemDataSource', fileSystemDataSource_fileSystemType - The file system type.

$sel:directoryPath:FileSystemDataSource', fileSystemDataSource_directoryPath - The full path to the directory to associate with the channel.

fileSystemDataSource_fileSystemId :: Lens' FileSystemDataSource Text

The file system id.

fileSystemDataSource_fileSystemAccessMode :: Lens' FileSystemDataSource FileSystemAccessMode

The access mode of the mount of the directory associated with the channel. A directory can be mounted either in ro (read-only) or rw (read-write) mode.

fileSystemDataSource_fileSystemType :: Lens' FileSystemDataSource FileSystemType

The file system type.

fileSystemDataSource_directoryPath :: Lens' FileSystemDataSource Text

The full path to the directory to associate with the channel.