Amazonka.SageMaker.Types.HyperParameterTuningJobSummary

Description

Documentation

data HyperParameterTuningJobSummary

Provides summary information about a hyperparameter tuning job.

See: newHyperParameterTuningJobSummary smart constructor.

Constructors

HyperParameterTuningJobSummary'

Fields

resourceLimits :: Maybe ResourceLimits

The ResourceLimits object that specifies the maximum number of training jobs and parallel training jobs allowed for this tuning job.

lastModifiedTime :: Maybe POSIX

The date and time that the tuning job was modified.

hyperParameterTuningEndTime :: Maybe POSIX

The date and time that the tuning job ended.

hyperParameterTuningJobName :: Text

The name of the tuning job.

hyperParameterTuningJobArn :: Text

The Amazon Resource Name (ARN) of the tuning job.

hyperParameterTuningJobStatus :: HyperParameterTuningJobStatus

The status of the tuning job.

strategy :: HyperParameterTuningJobStrategyType

Specifies the search strategy hyperparameter tuning uses to choose which hyperparameters to use for each iteration. Currently, the only valid value is Bayesian.

creationTime :: POSIX

The date and time that the tuning job was created.

trainingJobStatusCounters :: TrainingJobStatusCounters

The TrainingJobStatusCounters object that specifies the numbers of training jobs, categorized by status, that this tuning job launched.

objectiveStatusCounters :: ObjectiveStatusCounters

The ObjectiveStatusCounters object that specifies the numbers of training jobs, categorized by objective metric status, that this tuning job launched.

newHyperParameterTuningJobSummary

Arguments

:: Text
$sel:hyperParameterTuningJobName:HyperParameterTuningJobSummary'

-> Text
$sel:hyperParameterTuningJobArn:HyperParameterTuningJobSummary'

-> HyperParameterTuningJobStatus
$sel:hyperParameterTuningJobStatus:HyperParameterTuningJobSummary'

-> HyperParameterTuningJobStrategyType
$sel:strategy:HyperParameterTuningJobSummary'

-> UTCTime
$sel:creationTime:HyperParameterTuningJobSummary'

-> TrainingJobStatusCounters
$sel:trainingJobStatusCounters:HyperParameterTuningJobSummary'

-> ObjectiveStatusCounters
$sel:objectiveStatusCounters:HyperParameterTuningJobSummary'

-> HyperParameterTuningJobSummary

Create a value of HyperParameterTuningJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceLimits:HyperParameterTuningJobSummary', hyperParameterTuningJobSummary_resourceLimits - The ResourceLimits object that specifies the maximum number of training jobs and parallel training jobs allowed for this tuning job.

$sel:lastModifiedTime:HyperParameterTuningJobSummary', hyperParameterTuningJobSummary_lastModifiedTime - The date and time that the tuning job was modified.

$sel:hyperParameterTuningEndTime:HyperParameterTuningJobSummary', hyperParameterTuningJobSummary_hyperParameterTuningEndTime - The date and time that the tuning job ended.

$sel:hyperParameterTuningJobName:HyperParameterTuningJobSummary', hyperParameterTuningJobSummary_hyperParameterTuningJobName - The name of the tuning job.

$sel:hyperParameterTuningJobArn:HyperParameterTuningJobSummary', hyperParameterTuningJobSummary_hyperParameterTuningJobArn - The Amazon Resource Name (ARN) of the tuning job.

$sel:hyperParameterTuningJobStatus:HyperParameterTuningJobSummary', hyperParameterTuningJobSummary_hyperParameterTuningJobStatus - The status of the tuning job.

$sel:strategy:HyperParameterTuningJobSummary', hyperParameterTuningJobSummary_strategy - Specifies the search strategy hyperparameter tuning uses to choose which hyperparameters to use for each iteration. Currently, the only valid value is Bayesian.

$sel:creationTime:HyperParameterTuningJobSummary', hyperParameterTuningJobSummary_creationTime - The date and time that the tuning job was created.

$sel:trainingJobStatusCounters:HyperParameterTuningJobSummary', hyperParameterTuningJobSummary_trainingJobStatusCounters - The TrainingJobStatusCounters object that specifies the numbers of training jobs, categorized by status, that this tuning job launched.

$sel:objectiveStatusCounters:HyperParameterTuningJobSummary', hyperParameterTuningJobSummary_objectiveStatusCounters - The ObjectiveStatusCounters object that specifies the numbers of training jobs, categorized by objective metric status, that this tuning job launched.

hyperParameterTuningJobSummary_resourceLimits :: Lens' HyperParameterTuningJobSummary (Maybe ResourceLimits)

The ResourceLimits object that specifies the maximum number of training jobs and parallel training jobs allowed for this tuning job.

hyperParameterTuningJobSummary_lastModifiedTime :: Lens' HyperParameterTuningJobSummary (Maybe UTCTime)

The date and time that the tuning job was modified.

hyperParameterTuningJobSummary_hyperParameterTuningEndTime :: Lens' HyperParameterTuningJobSummary (Maybe UTCTime)

The date and time that the tuning job ended.

hyperParameterTuningJobSummary_hyperParameterTuningJobName :: Lens' HyperParameterTuningJobSummary Text

The name of the tuning job.

hyperParameterTuningJobSummary_hyperParameterTuningJobArn :: Lens' HyperParameterTuningJobSummary Text

The Amazon Resource Name (ARN) of the tuning job.

hyperParameterTuningJobSummary_hyperParameterTuningJobStatus :: Lens' HyperParameterTuningJobSummary HyperParameterTuningJobStatus

The status of the tuning job.

hyperParameterTuningJobSummary_strategy :: Lens' HyperParameterTuningJobSummary HyperParameterTuningJobStrategyType

Specifies the search strategy hyperparameter tuning uses to choose which hyperparameters to use for each iteration. Currently, the only valid value is Bayesian.

hyperParameterTuningJobSummary_creationTime :: Lens' HyperParameterTuningJobSummary UTCTime

The date and time that the tuning job was created.

hyperParameterTuningJobSummary_trainingJobStatusCounters :: Lens' HyperParameterTuningJobSummary TrainingJobStatusCounters

The TrainingJobStatusCounters object that specifies the numbers of training jobs, categorized by status, that this tuning job launched.

hyperParameterTuningJobSummary_objectiveStatusCounters :: Lens' HyperParameterTuningJobSummary ObjectiveStatusCounters

The ObjectiveStatusCounters object that specifies the numbers of training jobs, categorized by objective metric status, that this tuning job launched.