Amazonka.SageMaker.Types.Image

Description

Documentation

data Image

A SageMaker image. A SageMaker image represents a set of container images that are derived from a common base container image. Each of these container images is represented by a SageMaker ImageVersion.

See: newImage smart constructor.

Constructors

Image'

Fields

failureReason :: Maybe Text

When a create, update, or delete operation fails, the reason for the failure.

displayName :: Maybe Text

The name of the image as displayed.

description :: Maybe Text

The description of the image.

creationTime :: POSIX

When the image was created.

imageArn :: Text

The Amazon Resource Name (ARN) of the image.

imageName :: Text

The name of the image.

imageStatus :: ImageStatus

The status of the image.

lastModifiedTime :: POSIX

When the image was last modified.

newImage

Arguments

:: UTCTime
$sel:creationTime:Image'

-> Text
$sel:imageArn:Image'

-> Text
$sel:imageName:Image'

-> ImageStatus
$sel:imageStatus:Image'

-> UTCTime
$sel:lastModifiedTime:Image'

-> Image

Create a value of Image with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:Image', image_failureReason - When a create, update, or delete operation fails, the reason for the failure.

$sel:displayName:Image', image_displayName - The name of the image as displayed.

$sel:description:Image', image_description - The description of the image.

$sel:creationTime:Image', image_creationTime - When the image was created.

$sel:imageArn:Image', image_imageArn - The Amazon Resource Name (ARN) of the image.

$sel:imageName:Image', image_imageName - The name of the image.

$sel:imageStatus:Image', image_imageStatus - The status of the image.

$sel:lastModifiedTime:Image', image_lastModifiedTime - When the image was last modified.

image_failureReason :: Lens' Image (Maybe Text)

When a create, update, or delete operation fails, the reason for the failure.

image_displayName :: Lens' Image (Maybe Text)

The name of the image as displayed.

image_description :: Lens' Image (Maybe Text)

The description of the image.

image_creationTime :: Lens' Image UTCTime

When the image was created.

image_imageArn :: Lens' Image Text

The Amazon Resource Name (ARN) of the image.

image_imageName :: Lens' Image Text

The name of the image.

image_imageStatus :: Lens' Image ImageStatus

The status of the image.

image_lastModifiedTime :: Lens' Image UTCTime

When the image was last modified.