Amazonka.SageMaker.Types.ImageConfig

Description

Documentation

data ImageConfig

Specifies whether the model container is in Amazon ECR or a private Docker registry accessible from your Amazon Virtual Private Cloud (VPC).

See: newImageConfig smart constructor.

Constructors

ImageConfig'

Fields

repositoryAuthConfig :: Maybe RepositoryAuthConfig

(Optional) Specifies an authentication configuration for the private docker registry where your model image is hosted. Specify a value for this property only if you specified Vpc as the value for the RepositoryAccessMode field, and the private Docker registry where the model image is hosted requires authentication.

repositoryAccessMode :: RepositoryAccessMode

Set this to one of the following values:

Platform - The model image is hosted in Amazon ECR.

Vpc - The model image is hosted in a private Docker registry in your VPC.

newImageConfig

Arguments

:: RepositoryAccessMode
$sel:repositoryAccessMode:ImageConfig'

-> ImageConfig

Create a value of ImageConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:repositoryAuthConfig:ImageConfig', imageConfig_repositoryAuthConfig - (Optional) Specifies an authentication configuration for the private docker registry where your model image is hosted. Specify a value for this property only if you specified Vpc as the value for the RepositoryAccessMode field, and the private Docker registry where the model image is hosted requires authentication.

$sel:repositoryAccessMode:ImageConfig', imageConfig_repositoryAccessMode - Set this to one of the following values:

Platform - The model image is hosted in Amazon ECR.

Vpc - The model image is hosted in a private Docker registry in your VPC.

imageConfig_repositoryAuthConfig :: Lens' ImageConfig (Maybe RepositoryAuthConfig)

(Optional) Specifies an authentication configuration for the private docker registry where your model image is hosted. Specify a value for this property only if you specified Vpc as the value for the RepositoryAccessMode field, and the private Docker registry where the model image is hosted requires authentication.

imageConfig_repositoryAccessMode :: Lens' ImageConfig RepositoryAccessMode

Set this to one of the following values:

Platform - The model image is hosted in Amazon ECR.

Vpc - The model image is hosted in a private Docker registry in your VPC.