Amazonka.SageMaker.Types.ImageVersion

Description

Documentation

data ImageVersion

A version of a SageMaker Image. A version represents an existing container image.

See: newImageVersion smart constructor.

Constructors

ImageVersion'

Fields

failureReason :: Maybe Text

When a create or delete operation fails, the reason for the failure.

creationTime :: POSIX

When the version was created.

imageArn :: Text

The Amazon Resource Name (ARN) of the image the version is based on.

imageVersionArn :: Text

The ARN of the version.

imageVersionStatus :: ImageVersionStatus

The status of the version.

lastModifiedTime :: POSIX

When the version was last modified.

version :: Natural

The version number.

newImageVersion

Arguments

:: UTCTime
$sel:creationTime:ImageVersion'

-> Text
$sel:imageArn:ImageVersion'

-> Text
$sel:imageVersionArn:ImageVersion'

-> ImageVersionStatus
$sel:imageVersionStatus:ImageVersion'

-> UTCTime
$sel:lastModifiedTime:ImageVersion'

-> Natural
$sel:version:ImageVersion'

-> ImageVersion

Create a value of ImageVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:ImageVersion', imageVersion_failureReason - When a create or delete operation fails, the reason for the failure.

$sel:creationTime:ImageVersion', imageVersion_creationTime - When the version was created.

$sel:imageArn:ImageVersion', imageVersion_imageArn - The Amazon Resource Name (ARN) of the image the version is based on.

$sel:imageVersionArn:ImageVersion', imageVersion_imageVersionArn - The ARN of the version.

$sel:imageVersionStatus:ImageVersion', imageVersion_imageVersionStatus - The status of the version.

$sel:lastModifiedTime:ImageVersion', imageVersion_lastModifiedTime - When the version was last modified.

$sel:version:ImageVersion', imageVersion_version - The version number.

imageVersion_failureReason :: Lens' ImageVersion (Maybe Text)

When a create or delete operation fails, the reason for the failure.

imageVersion_creationTime :: Lens' ImageVersion UTCTime

When the version was created.

imageVersion_imageArn :: Lens' ImageVersion Text

The Amazon Resource Name (ARN) of the image the version is based on.

imageVersion_imageVersionArn :: Lens' ImageVersion Text

The ARN of the version.

imageVersion_imageVersionStatus :: Lens' ImageVersion ImageVersionStatus

The status of the version.

imageVersion_lastModifiedTime :: Lens' ImageVersion UTCTime

When the version was last modified.

imageVersion_version :: Lens' ImageVersion Natural

The version number.