Amazonka.SageMaker.Types.InferenceSpecification

Description

Documentation

data InferenceSpecification

Defines how to perform inference generation after a training job is run.

See: newInferenceSpecification smart constructor.

Constructors

InferenceSpecification'

Fields

supportedRealtimeInferenceInstanceTypes :: Maybe [ProductionVariantInstanceType]

A list of the instance types that are used to generate inferences in real-time.

This parameter is required for unversioned models, and optional for versioned models.

supportedTransformInstanceTypes :: Maybe (NonEmpty TransformInstanceType)

A list of the instance types on which a transformation job can be run or on which an endpoint can be deployed.

This parameter is required for unversioned models, and optional for versioned models.

containers :: NonEmpty ModelPackageContainerDefinition

The Amazon ECR registry path of the Docker image that contains the inference code.

supportedContentTypes :: [Text]

The supported MIME types for the input data.

supportedResponseMIMETypes :: [Text]

The supported MIME types for the output data.

newInferenceSpecification

Arguments

:: NonEmpty ModelPackageContainerDefinition
$sel:containers:InferenceSpecification'

-> InferenceSpecification

Create a value of InferenceSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:supportedRealtimeInferenceInstanceTypes:InferenceSpecification', inferenceSpecification_supportedRealtimeInferenceInstanceTypes - A list of the instance types that are used to generate inferences in real-time.

This parameter is required for unversioned models, and optional for versioned models.

$sel:supportedTransformInstanceTypes:InferenceSpecification', inferenceSpecification_supportedTransformInstanceTypes - A list of the instance types on which a transformation job can be run or on which an endpoint can be deployed.

This parameter is required for unversioned models, and optional for versioned models.

$sel:containers:InferenceSpecification', inferenceSpecification_containers - The Amazon ECR registry path of the Docker image that contains the inference code.

$sel:supportedContentTypes:InferenceSpecification', inferenceSpecification_supportedContentTypes - The supported MIME types for the input data.

$sel:supportedResponseMIMETypes:InferenceSpecification', inferenceSpecification_supportedResponseMIMETypes - The supported MIME types for the output data.

inferenceSpecification_supportedRealtimeInferenceInstanceTypes :: Lens' InferenceSpecification (Maybe [ProductionVariantInstanceType])

A list of the instance types that are used to generate inferences in real-time.

This parameter is required for unversioned models, and optional for versioned models.

inferenceSpecification_supportedTransformInstanceTypes :: Lens' InferenceSpecification (Maybe (NonEmpty TransformInstanceType))

A list of the instance types on which a transformation job can be run or on which an endpoint can be deployed.

This parameter is required for unversioned models, and optional for versioned models.

inferenceSpecification_containers :: Lens' InferenceSpecification (NonEmpty ModelPackageContainerDefinition)

The Amazon ECR registry path of the Docker image that contains the inference code.

inferenceSpecification_supportedContentTypes :: Lens' InferenceSpecification [Text]

The supported MIME types for the input data.

inferenceSpecification_supportedResponseMIMETypes :: Lens' InferenceSpecification [Text]

The supported MIME types for the output data.