Amazonka.SageMaker.Types.ModelBiasAppSpecification

Description

Documentation

data ModelBiasAppSpecification

Docker container image configuration object for the model bias job.

See: newModelBiasAppSpecification smart constructor.

Constructors

ModelBiasAppSpecification'

Fields

environment :: Maybe (HashMap Text Text)

Sets the environment variables in the Docker container.

imageUri :: Text

The container image to be run by the model bias job.

configUri :: Text

JSON formatted S3 file that defines bias parameters. For more information on this JSON configuration file, see Configure bias parameters.

newModelBiasAppSpecification

Arguments

:: Text
$sel:imageUri:ModelBiasAppSpecification'

-> Text
$sel:configUri:ModelBiasAppSpecification'

-> ModelBiasAppSpecification

Create a value of ModelBiasAppSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environment:ModelBiasAppSpecification', modelBiasAppSpecification_environment - Sets the environment variables in the Docker container.

$sel:imageUri:ModelBiasAppSpecification', modelBiasAppSpecification_imageUri - The container image to be run by the model bias job.

$sel:configUri:ModelBiasAppSpecification', modelBiasAppSpecification_configUri - JSON formatted S3 file that defines bias parameters. For more information on this JSON configuration file, see Configure bias parameters.

modelBiasAppSpecification_environment :: Lens' ModelBiasAppSpecification (Maybe (HashMap Text Text))

Sets the environment variables in the Docker container.

modelBiasAppSpecification_imageUri :: Lens' ModelBiasAppSpecification Text

The container image to be run by the model bias job.

modelBiasAppSpecification_configUri :: Lens' ModelBiasAppSpecification Text

JSON formatted S3 file that defines bias parameters. For more information on this JSON configuration file, see Configure bias parameters.