Amazonka.SageMaker.Types.ModelPackageStatusDetails

Description

Documentation

data ModelPackageStatusDetails

Specifies the validation and image scan statuses of the model package.

See: newModelPackageStatusDetails smart constructor.

Constructors

ModelPackageStatusDetails'

Fields

imageScanStatuses :: Maybe [ModelPackageStatusItem]

The status of the scan of the Docker image container for the model package.

validationStatuses :: [ModelPackageStatusItem]

The validation status of the model package.

newModelPackageStatusDetails :: ModelPackageStatusDetails

Create a value of ModelPackageStatusDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageScanStatuses:ModelPackageStatusDetails', modelPackageStatusDetails_imageScanStatuses - The status of the scan of the Docker image container for the model package.

$sel:validationStatuses:ModelPackageStatusDetails', modelPackageStatusDetails_validationStatuses - The validation status of the model package.

modelPackageStatusDetails_imageScanStatuses :: Lens' ModelPackageStatusDetails (Maybe [ModelPackageStatusItem])

The status of the scan of the Docker image container for the model package.

modelPackageStatusDetails_validationStatuses :: Lens' ModelPackageStatusDetails [ModelPackageStatusItem]

The validation status of the model package.