Amazonka.SageMaker.Types.ModelPackageSummary

Description

Documentation

data ModelPackageSummary

Provides summary information about a model package.

See: newModelPackageSummary smart constructor.

Constructors

ModelPackageSummary'

Fields

modelApprovalStatus :: Maybe ModelApprovalStatus

The approval status of the model. This can be one of the following values.

APPROVED - The model is approved

REJECTED - The model is rejected.

PENDING_MANUAL_APPROVAL - The model is waiting for manual approval.

modelPackageDescription :: Maybe Text

A brief description of the model package.

modelPackageVersion :: Maybe Natural

If the model package is a versioned model, the version of the model.

modelPackageGroupName :: Maybe Text

If the model package is a versioned model, the model group that the versioned model belongs to.

modelPackageName :: Text

The name of the model package.

modelPackageArn :: Text

The Amazon Resource Name (ARN) of the model package.

creationTime :: POSIX

A timestamp that shows when the model package was created.

modelPackageStatus :: ModelPackageStatus

The overall status of the model package.

newModelPackageSummary

Arguments

:: Text
$sel:modelPackageName:ModelPackageSummary'

-> Text
$sel:modelPackageArn:ModelPackageSummary'

-> UTCTime
$sel:creationTime:ModelPackageSummary'

-> ModelPackageStatus
$sel:modelPackageStatus:ModelPackageSummary'

-> ModelPackageSummary

Create a value of ModelPackageSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelApprovalStatus:ModelPackageSummary', modelPackageSummary_modelApprovalStatus - The approval status of the model. This can be one of the following values.

APPROVED - The model is approved

REJECTED - The model is rejected.

PENDING_MANUAL_APPROVAL - The model is waiting for manual approval.

$sel:modelPackageDescription:ModelPackageSummary', modelPackageSummary_modelPackageDescription - A brief description of the model package.

$sel:modelPackageVersion:ModelPackageSummary', modelPackageSummary_modelPackageVersion - If the model package is a versioned model, the version of the model.

$sel:modelPackageGroupName:ModelPackageSummary', modelPackageSummary_modelPackageGroupName - If the model package is a versioned model, the model group that the versioned model belongs to.

$sel:modelPackageName:ModelPackageSummary', modelPackageSummary_modelPackageName - The name of the model package.

$sel:modelPackageArn:ModelPackageSummary', modelPackageSummary_modelPackageArn - The Amazon Resource Name (ARN) of the model package.

$sel:creationTime:ModelPackageSummary', modelPackageSummary_creationTime - A timestamp that shows when the model package was created.

$sel:modelPackageStatus:ModelPackageSummary', modelPackageSummary_modelPackageStatus - The overall status of the model package.

modelPackageSummary_modelApprovalStatus :: Lens' ModelPackageSummary (Maybe ModelApprovalStatus)

The approval status of the model. This can be one of the following values.

APPROVED - The model is approved

REJECTED - The model is rejected.

PENDING_MANUAL_APPROVAL - The model is waiting for manual approval.

modelPackageSummary_modelPackageDescription :: Lens' ModelPackageSummary (Maybe Text)

A brief description of the model package.

modelPackageSummary_modelPackageVersion :: Lens' ModelPackageSummary (Maybe Natural)

If the model package is a versioned model, the version of the model.

modelPackageSummary_modelPackageGroupName :: Lens' ModelPackageSummary (Maybe Text)

If the model package is a versioned model, the model group that the versioned model belongs to.

modelPackageSummary_modelPackageName :: Lens' ModelPackageSummary Text

The name of the model package.

modelPackageSummary_modelPackageArn :: Lens' ModelPackageSummary Text

The Amazon Resource Name (ARN) of the model package.

modelPackageSummary_creationTime :: Lens' ModelPackageSummary UTCTime

A timestamp that shows when the model package was created.

modelPackageSummary_modelPackageStatus :: Lens' ModelPackageSummary ModelPackageStatus

The overall status of the model package.