Amazonka.SageMaker.Types.MonitoringClusterConfig

Description

Documentation

data MonitoringClusterConfig

Configuration for the cluster used to run model monitoring jobs.

See: newMonitoringClusterConfig smart constructor.

Constructors

MonitoringClusterConfig'

Fields

volumeKmsKeyId :: Maybe Text

The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data on the storage volume attached to the ML compute instance(s) that run the model monitoring job.

instanceCount :: Natural

The number of ML compute instances to use in the model monitoring job. For distributed processing jobs, specify a value greater than 1. The default value is 1.

instanceType :: ProcessingInstanceType

The ML compute instance type for the processing job.

volumeSizeInGB :: Natural

The size of the ML storage volume, in gigabytes, that you want to provision. You must specify sufficient ML storage for your scenario.

newMonitoringClusterConfig

Arguments

:: Natural
$sel:instanceCount:MonitoringClusterConfig'

-> ProcessingInstanceType
$sel:instanceType:MonitoringClusterConfig'

-> Natural
$sel:volumeSizeInGB:MonitoringClusterConfig'

-> MonitoringClusterConfig

Create a value of MonitoringClusterConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeKmsKeyId:MonitoringClusterConfig', monitoringClusterConfig_volumeKmsKeyId - The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data on the storage volume attached to the ML compute instance(s) that run the model monitoring job.

$sel:instanceCount:MonitoringClusterConfig', monitoringClusterConfig_instanceCount - The number of ML compute instances to use in the model monitoring job. For distributed processing jobs, specify a value greater than 1. The default value is 1.

$sel:instanceType:MonitoringClusterConfig', monitoringClusterConfig_instanceType - The ML compute instance type for the processing job.

$sel:volumeSizeInGB:MonitoringClusterConfig', monitoringClusterConfig_volumeSizeInGB - The size of the ML storage volume, in gigabytes, that you want to provision. You must specify sufficient ML storage for your scenario.

monitoringClusterConfig_volumeKmsKeyId :: Lens' MonitoringClusterConfig (Maybe Text)

The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt data on the storage volume attached to the ML compute instance(s) that run the model monitoring job.

monitoringClusterConfig_instanceCount :: Lens' MonitoringClusterConfig Natural

The number of ML compute instances to use in the model monitoring job. For distributed processing jobs, specify a value greater than 1. The default value is 1.

monitoringClusterConfig_instanceType :: Lens' MonitoringClusterConfig ProcessingInstanceType

The ML compute instance type for the processing job.

monitoringClusterConfig_volumeSizeInGB :: Lens' MonitoringClusterConfig Natural

The size of the ML storage volume, in gigabytes, that you want to provision. You must specify sufficient ML storage for your scenario.