Amazonka.SageMaker.Types.MonitoringExecutionSummary

Description

Documentation

data MonitoringExecutionSummary

Summary of information about the last monitoring job to run.

See: newMonitoringExecutionSummary smart constructor.

Constructors

MonitoringExecutionSummary'

Fields

monitoringType :: Maybe MonitoringType

The type of the monitoring job.

failureReason :: Maybe Text

Contains the reason a monitoring job failed, if it failed.

endpointName :: Maybe Text

The name of the endpoint used to run the monitoring job.

processingJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the monitoring job.

monitoringJobDefinitionName :: Maybe Text

The name of the monitoring job.

monitoringScheduleName :: Text

The name of the monitoring schedule.

scheduledTime :: POSIX

The time the monitoring job was scheduled.

creationTime :: POSIX

The time at which the monitoring job was created.

lastModifiedTime :: POSIX

A timestamp that indicates the last time the monitoring job was modified.

monitoringExecutionStatus :: ExecutionStatus

The status of the monitoring job.

newMonitoringExecutionSummary

Arguments

:: Text
$sel:monitoringScheduleName:MonitoringExecutionSummary'

-> UTCTime
$sel:scheduledTime:MonitoringExecutionSummary'

-> UTCTime
$sel:creationTime:MonitoringExecutionSummary'

-> UTCTime
$sel:lastModifiedTime:MonitoringExecutionSummary'

-> ExecutionStatus
$sel:monitoringExecutionStatus:MonitoringExecutionSummary'

-> MonitoringExecutionSummary

Create a value of MonitoringExecutionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monitoringType:MonitoringExecutionSummary', monitoringExecutionSummary_monitoringType - The type of the monitoring job.

$sel:failureReason:MonitoringExecutionSummary', monitoringExecutionSummary_failureReason - Contains the reason a monitoring job failed, if it failed.

$sel:endpointName:MonitoringExecutionSummary', monitoringExecutionSummary_endpointName - The name of the endpoint used to run the monitoring job.

$sel:processingJobArn:MonitoringExecutionSummary', monitoringExecutionSummary_processingJobArn - The Amazon Resource Name (ARN) of the monitoring job.

$sel:monitoringJobDefinitionName:MonitoringExecutionSummary', monitoringExecutionSummary_monitoringJobDefinitionName - The name of the monitoring job.

$sel:monitoringScheduleName:MonitoringExecutionSummary', monitoringExecutionSummary_monitoringScheduleName - The name of the monitoring schedule.

$sel:scheduledTime:MonitoringExecutionSummary', monitoringExecutionSummary_scheduledTime - The time the monitoring job was scheduled.

$sel:creationTime:MonitoringExecutionSummary', monitoringExecutionSummary_creationTime - The time at which the monitoring job was created.

$sel:lastModifiedTime:MonitoringExecutionSummary', monitoringExecutionSummary_lastModifiedTime - A timestamp that indicates the last time the monitoring job was modified.

$sel:monitoringExecutionStatus:MonitoringExecutionSummary', monitoringExecutionSummary_monitoringExecutionStatus - The status of the monitoring job.

monitoringExecutionSummary_monitoringType :: Lens' MonitoringExecutionSummary (Maybe MonitoringType)

The type of the monitoring job.

monitoringExecutionSummary_failureReason :: Lens' MonitoringExecutionSummary (Maybe Text)

Contains the reason a monitoring job failed, if it failed.

monitoringExecutionSummary_endpointName :: Lens' MonitoringExecutionSummary (Maybe Text)

The name of the endpoint used to run the monitoring job.

monitoringExecutionSummary_processingJobArn :: Lens' MonitoringExecutionSummary (Maybe Text)

The Amazon Resource Name (ARN) of the monitoring job.

monitoringExecutionSummary_monitoringJobDefinitionName :: Lens' MonitoringExecutionSummary (Maybe Text)

The name of the monitoring job.

monitoringExecutionSummary_monitoringScheduleName :: Lens' MonitoringExecutionSummary Text

The name of the monitoring schedule.

monitoringExecutionSummary_scheduledTime :: Lens' MonitoringExecutionSummary UTCTime

The time the monitoring job was scheduled.

monitoringExecutionSummary_creationTime :: Lens' MonitoringExecutionSummary UTCTime

The time at which the monitoring job was created.

monitoringExecutionSummary_lastModifiedTime :: Lens' MonitoringExecutionSummary UTCTime

A timestamp that indicates the last time the monitoring job was modified.

monitoringExecutionSummary_monitoringExecutionStatus :: Lens' MonitoringExecutionSummary ExecutionStatus

The status of the monitoring job.