Amazonka.SageMaker.Types.MonitoringSchedule

Description

Documentation

data MonitoringSchedule

A schedule for a model monitoring job. For information about model monitor, see Amazon SageMaker Model Monitor.

See: newMonitoringSchedule smart constructor.

Constructors

MonitoringSchedule'

Fields

creationTime :: Maybe POSIX

The time that the monitoring schedule was created.

monitoringType :: Maybe MonitoringType

The type of the monitoring job definition to schedule.

failureReason :: Maybe Text

If the monitoring schedule failed, the reason it failed.

monitoringScheduleArn :: Maybe Text

The Amazon Resource Name (ARN) of the monitoring schedule.

endpointName :: Maybe Text

The endpoint that hosts the model being monitored.

lastModifiedTime :: Maybe POSIX

The last time the monitoring schedule was changed.

monitoringScheduleStatus :: Maybe ScheduleStatus

The status of the monitoring schedule. This can be one of the following values.

PENDING - The schedule is pending being created.

FAILED - The schedule failed.

SCHEDULED - The schedule was successfully created.

STOPPED - The schedule was stopped.

lastMonitoringExecutionSummary :: Maybe MonitoringExecutionSummary

monitoringScheduleConfig :: Maybe MonitoringScheduleConfig

monitoringScheduleName :: Maybe Text

The name of the monitoring schedule.

tags :: Maybe [Tag]

A list of the tags associated with the monitoring schedlue. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

newMonitoringSchedule :: MonitoringSchedule

Create a value of MonitoringSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:MonitoringSchedule', monitoringSchedule_creationTime - The time that the monitoring schedule was created.

$sel:monitoringType:MonitoringSchedule', monitoringSchedule_monitoringType - The type of the monitoring job definition to schedule.

$sel:failureReason:MonitoringSchedule', monitoringSchedule_failureReason - If the monitoring schedule failed, the reason it failed.

$sel:monitoringScheduleArn:MonitoringSchedule', monitoringSchedule_monitoringScheduleArn - The Amazon Resource Name (ARN) of the monitoring schedule.

$sel:endpointName:MonitoringSchedule', monitoringSchedule_endpointName - The endpoint that hosts the model being monitored.

$sel:lastModifiedTime:MonitoringSchedule', monitoringSchedule_lastModifiedTime - The last time the monitoring schedule was changed.

$sel:monitoringScheduleStatus:MonitoringSchedule', monitoringSchedule_monitoringScheduleStatus - The status of the monitoring schedule. This can be one of the following values.

PENDING - The schedule is pending being created.

FAILED - The schedule failed.

SCHEDULED - The schedule was successfully created.

STOPPED - The schedule was stopped.

$sel:lastMonitoringExecutionSummary:MonitoringSchedule', monitoringSchedule_lastMonitoringExecutionSummary - Undocumented member.

$sel:monitoringScheduleConfig:MonitoringSchedule', monitoringSchedule_monitoringScheduleConfig - Undocumented member.

$sel:monitoringScheduleName:MonitoringSchedule', monitoringSchedule_monitoringScheduleName - The name of the monitoring schedule.

$sel:tags:MonitoringSchedule', monitoringSchedule_tags - A list of the tags associated with the monitoring schedlue. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.

monitoringSchedule_creationTime :: Lens' MonitoringSchedule (Maybe UTCTime)

The time that the monitoring schedule was created.

monitoringSchedule_monitoringType :: Lens' MonitoringSchedule (Maybe MonitoringType)

The type of the monitoring job definition to schedule.

monitoringSchedule_failureReason :: Lens' MonitoringSchedule (Maybe Text)

If the monitoring schedule failed, the reason it failed.

monitoringSchedule_monitoringScheduleArn :: Lens' MonitoringSchedule (Maybe Text)

The Amazon Resource Name (ARN) of the monitoring schedule.

monitoringSchedule_endpointName :: Lens' MonitoringSchedule (Maybe Text)

The endpoint that hosts the model being monitored.

monitoringSchedule_lastModifiedTime :: Lens' MonitoringSchedule (Maybe UTCTime)

The last time the monitoring schedule was changed.

monitoringSchedule_monitoringScheduleStatus :: Lens' MonitoringSchedule (Maybe ScheduleStatus)

The status of the monitoring schedule. This can be one of the following values.

PENDING - The schedule is pending being created.

FAILED - The schedule failed.

SCHEDULED - The schedule was successfully created.

STOPPED - The schedule was stopped.

monitoringSchedule_lastMonitoringExecutionSummary :: Lens' MonitoringSchedule (Maybe MonitoringExecutionSummary)

Undocumented member.

monitoringSchedule_monitoringScheduleConfig :: Lens' MonitoringSchedule (Maybe MonitoringScheduleConfig)

Undocumented member.

monitoringSchedule_monitoringScheduleName :: Lens' MonitoringSchedule (Maybe Text)

The name of the monitoring schedule.

monitoringSchedule_tags :: Lens' MonitoringSchedule (Maybe [Tag])

A list of the tags associated with the monitoring schedlue. For more information, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference Guide.