Amazonka.SageMaker.Types.NotebookInstanceSummary

Description

Documentation

data NotebookInstanceSummary

Provides summary information for an Amazon SageMaker notebook instance.

See: newNotebookInstanceSummary smart constructor.

Constructors

NotebookInstanceSummary'

Fields

creationTime :: Maybe POSIX

A timestamp that shows when the notebook instance was created.

additionalCodeRepositories :: Maybe [Text]

An array of up to three Git repositories associated with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

url :: Maybe Text

The URL that you use to connect to the Jupyter instance running in your notebook instance.

lastModifiedTime :: Maybe POSIX

A timestamp that shows when the notebook instance was last modified.

instanceType :: Maybe InstanceType

The type of ML compute instance that the notebook instance is running on.

notebookInstanceStatus :: Maybe NotebookInstanceStatus

The status of the notebook instance.

defaultCodeRepository :: Maybe Text

The Git repository associated with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

notebookInstanceLifecycleConfigName :: Maybe Text

The name of a notebook instance lifecycle configuration associated with this notebook instance.

For information about notebook instance lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

notebookInstanceName :: Text

The name of the notebook instance that you want a summary for.

notebookInstanceArn :: Text

The Amazon Resource Name (ARN) of the notebook instance.

newNotebookInstanceSummary

Arguments

:: Text
$sel:notebookInstanceName:NotebookInstanceSummary'

-> Text
$sel:notebookInstanceArn:NotebookInstanceSummary'

-> NotebookInstanceSummary

Create a value of NotebookInstanceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:NotebookInstanceSummary', notebookInstanceSummary_creationTime - A timestamp that shows when the notebook instance was created.

$sel:additionalCodeRepositories:NotebookInstanceSummary', notebookInstanceSummary_additionalCodeRepositories - An array of up to three Git repositories associated with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

$sel:url:NotebookInstanceSummary', notebookInstanceSummary_url - The URL that you use to connect to the Jupyter instance running in your notebook instance.

$sel:lastModifiedTime:NotebookInstanceSummary', notebookInstanceSummary_lastModifiedTime - A timestamp that shows when the notebook instance was last modified.

$sel:instanceType:NotebookInstanceSummary', notebookInstanceSummary_instanceType - The type of ML compute instance that the notebook instance is running on.

$sel:notebookInstanceStatus:NotebookInstanceSummary', notebookInstanceSummary_notebookInstanceStatus - The status of the notebook instance.

$sel:defaultCodeRepository:NotebookInstanceSummary', notebookInstanceSummary_defaultCodeRepository - The Git repository associated with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

$sel:notebookInstanceLifecycleConfigName:NotebookInstanceSummary', notebookInstanceSummary_notebookInstanceLifecycleConfigName - The name of a notebook instance lifecycle configuration associated with this notebook instance.

For information about notebook instance lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

$sel:notebookInstanceName:NotebookInstanceSummary', notebookInstanceSummary_notebookInstanceName - The name of the notebook instance that you want a summary for.

$sel:notebookInstanceArn:NotebookInstanceSummary', notebookInstanceSummary_notebookInstanceArn - The Amazon Resource Name (ARN) of the notebook instance.

notebookInstanceSummary_creationTime :: Lens' NotebookInstanceSummary (Maybe UTCTime)

A timestamp that shows when the notebook instance was created.

notebookInstanceSummary_additionalCodeRepositories :: Lens' NotebookInstanceSummary (Maybe [Text])

An array of up to three Git repositories associated with the notebook instance. These can be either the names of Git repositories stored as resources in your account, or the URL of Git repositories in Amazon Web Services CodeCommit or in any other Git repository. These repositories are cloned at the same level as the default repository of your notebook instance. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

notebookInstanceSummary_url :: Lens' NotebookInstanceSummary (Maybe Text)

The URL that you use to connect to the Jupyter instance running in your notebook instance.

notebookInstanceSummary_lastModifiedTime :: Lens' NotebookInstanceSummary (Maybe UTCTime)

A timestamp that shows when the notebook instance was last modified.

notebookInstanceSummary_instanceType :: Lens' NotebookInstanceSummary (Maybe InstanceType)

The type of ML compute instance that the notebook instance is running on.

notebookInstanceSummary_notebookInstanceStatus :: Lens' NotebookInstanceSummary (Maybe NotebookInstanceStatus)

The status of the notebook instance.

notebookInstanceSummary_defaultCodeRepository :: Lens' NotebookInstanceSummary (Maybe Text)

The Git repository associated with the notebook instance as its default code repository. This can be either the name of a Git repository stored as a resource in your account, or the URL of a Git repository in Amazon Web Services CodeCommit or in any other Git repository. When you open a notebook instance, it opens in the directory that contains this repository. For more information, see Associating Git Repositories with Amazon SageMaker Notebook Instances.

notebookInstanceSummary_notebookInstanceLifecycleConfigName :: Lens' NotebookInstanceSummary (Maybe Text)

The name of a notebook instance lifecycle configuration associated with this notebook instance.

For information about notebook instance lifestyle configurations, see Step 2.1: (Optional) Customize a Notebook Instance.

notebookInstanceSummary_notebookInstanceName :: Lens' NotebookInstanceSummary Text

The name of the notebook instance that you want a summary for.

notebookInstanceSummary_notebookInstanceArn :: Lens' NotebookInstanceSummary Text

The Amazon Resource Name (ARN) of the notebook instance.