Amazonka.SageMaker.Types.OidcConfig

Description

Documentation

data OidcConfig

Use this parameter to configure your OIDC Identity Provider (IdP).

See: newOidcConfig smart constructor.

Constructors

OidcConfig'

Fields

clientId :: Text

The OIDC IdP client ID used to configure your private workforce.

clientSecret :: Sensitive Text

The OIDC IdP client secret used to configure your private workforce.

issuer :: Text

The OIDC IdP issuer used to configure your private workforce.

authorizationEndpoint :: Text

The OIDC IdP authorization endpoint used to configure your private workforce.

tokenEndpoint :: Text

The OIDC IdP token endpoint used to configure your private workforce.

userInfoEndpoint :: Text

The OIDC IdP user information endpoint used to configure your private workforce.

logoutEndpoint :: Text

The OIDC IdP logout endpoint used to configure your private workforce.

jwksUri :: Text

The OIDC IdP JSON Web Key Set (Jwks) URI used to configure your private workforce.

newOidcConfig

Arguments

:: Text
$sel:clientId:OidcConfig'

-> Text
$sel:clientSecret:OidcConfig'

-> Text
$sel:issuer:OidcConfig'

-> Text
$sel:authorizationEndpoint:OidcConfig'

-> Text
$sel:tokenEndpoint:OidcConfig'

-> Text
$sel:userInfoEndpoint:OidcConfig'

-> Text
$sel:logoutEndpoint:OidcConfig'

-> Text
$sel:jwksUri:OidcConfig'

-> OidcConfig

Create a value of OidcConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:OidcConfig', oidcConfig_clientId - The OIDC IdP client ID used to configure your private workforce.

$sel:clientSecret:OidcConfig', oidcConfig_clientSecret - The OIDC IdP client secret used to configure your private workforce.

$sel:issuer:OidcConfig', oidcConfig_issuer - The OIDC IdP issuer used to configure your private workforce.

$sel:authorizationEndpoint:OidcConfig', oidcConfig_authorizationEndpoint - The OIDC IdP authorization endpoint used to configure your private workforce.

$sel:tokenEndpoint:OidcConfig', oidcConfig_tokenEndpoint - The OIDC IdP token endpoint used to configure your private workforce.

$sel:userInfoEndpoint:OidcConfig', oidcConfig_userInfoEndpoint - The OIDC IdP user information endpoint used to configure your private workforce.

$sel:logoutEndpoint:OidcConfig', oidcConfig_logoutEndpoint - The OIDC IdP logout endpoint used to configure your private workforce.

$sel:jwksUri:OidcConfig', oidcConfig_jwksUri - The OIDC IdP JSON Web Key Set (Jwks) URI used to configure your private workforce.

oidcConfig_clientId :: Lens' OidcConfig Text

The OIDC IdP client ID used to configure your private workforce.

oidcConfig_clientSecret :: Lens' OidcConfig Text

The OIDC IdP client secret used to configure your private workforce.

oidcConfig_issuer :: Lens' OidcConfig Text

The OIDC IdP issuer used to configure your private workforce.

oidcConfig_authorizationEndpoint :: Lens' OidcConfig Text

The OIDC IdP authorization endpoint used to configure your private workforce.

oidcConfig_tokenEndpoint :: Lens' OidcConfig Text

The OIDC IdP token endpoint used to configure your private workforce.

oidcConfig_userInfoEndpoint :: Lens' OidcConfig Text

The OIDC IdP user information endpoint used to configure your private workforce.

oidcConfig_logoutEndpoint :: Lens' OidcConfig Text

The OIDC IdP logout endpoint used to configure your private workforce.

oidcConfig_jwksUri :: Lens' OidcConfig Text

The OIDC IdP JSON Web Key Set (Jwks) URI used to configure your private workforce.