Amazonka.SageMaker.Types.OidcMemberDefinition

Description

Documentation

data OidcMemberDefinition

A list of user groups that exist in your OIDC Identity Provider (IdP). One to ten groups can be used to create a single private work team. When you add a user group to the list of Groups, you can add that user group to one or more private work teams. If you add a user group to a private work team, all workers in that user group are added to the work team.

See: newOidcMemberDefinition smart constructor.

Constructors

OidcMemberDefinition'

Fields

groups :: NonEmpty Text

A list of comma seperated strings that identifies user groups in your OIDC IdP. Each user group is made up of a group of private workers.

newOidcMemberDefinition

Arguments

:: NonEmpty Text
$sel:groups:OidcMemberDefinition'

-> OidcMemberDefinition

Create a value of OidcMemberDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:OidcMemberDefinition', oidcMemberDefinition_groups - A list of comma seperated strings that identifies user groups in your OIDC IdP. Each user group is made up of a group of private workers.

oidcMemberDefinition_groups :: Lens' OidcMemberDefinition (NonEmpty Text)

A list of comma seperated strings that identifies user groups in your OIDC IdP. Each user group is made up of a group of private workers.