Amazonka.SageMaker.Types.Parameter

Description

Documentation

data Parameter

Assigns a value to a named Pipeline parameter.

See: newParameter smart constructor.

Constructors

Parameter'

Fields

name :: Text

The name of the parameter to assign a value to. This parameter name must match a named parameter in the pipeline definition.

value :: Text

The literal value for the parameter.

newParameter

Arguments

:: Text
$sel:name:Parameter'

-> Text
$sel:value:Parameter'

-> Parameter

Create a value of Parameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Parameter', parameter_name - The name of the parameter to assign a value to. This parameter name must match a named parameter in the pipeline definition.

$sel:value:Parameter', parameter_value - The literal value for the parameter.

parameter_name :: Lens' Parameter Text

The name of the parameter to assign a value to. This parameter name must match a named parameter in the pipeline definition.

parameter_value :: Lens' Parameter Text

The literal value for the parameter.