Amazonka.SageMaker.Types.ParameterRanges

Description

Documentation

data ParameterRanges

Specifies ranges of integer, continuous, and categorical hyperparameters that a hyperparameter tuning job searches. The hyperparameter tuning job launches training jobs with hyperparameter values within these ranges to find the combination of values that result in the training job with the best performance as measured by the objective metric of the hyperparameter tuning job.

You can specify a maximum of 20 hyperparameters that a hyperparameter tuning job can search over. Every possible value of a categorical parameter range counts against this limit.

See: newParameterRanges smart constructor.

Constructors

ParameterRanges'

Fields

categoricalParameterRanges :: Maybe [CategoricalParameterRange]

The array of CategoricalParameterRange objects that specify ranges of categorical hyperparameters that a hyperparameter tuning job searches.

integerParameterRanges :: Maybe [IntegerParameterRange]

The array of IntegerParameterRange objects that specify ranges of integer hyperparameters that a hyperparameter tuning job searches.

continuousParameterRanges :: Maybe [ContinuousParameterRange]

The array of ContinuousParameterRange objects that specify ranges of continuous hyperparameters that a hyperparameter tuning job searches.

newParameterRanges :: ParameterRanges

Create a value of ParameterRanges with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:categoricalParameterRanges:ParameterRanges', parameterRanges_categoricalParameterRanges - The array of CategoricalParameterRange objects that specify ranges of categorical hyperparameters that a hyperparameter tuning job searches.

$sel:integerParameterRanges:ParameterRanges', parameterRanges_integerParameterRanges - The array of IntegerParameterRange objects that specify ranges of integer hyperparameters that a hyperparameter tuning job searches.

$sel:continuousParameterRanges:ParameterRanges', parameterRanges_continuousParameterRanges - The array of ContinuousParameterRange objects that specify ranges of continuous hyperparameters that a hyperparameter tuning job searches.

parameterRanges_categoricalParameterRanges :: Lens' ParameterRanges (Maybe [CategoricalParameterRange])

The array of CategoricalParameterRange objects that specify ranges of categorical hyperparameters that a hyperparameter tuning job searches.

parameterRanges_integerParameterRanges :: Lens' ParameterRanges (Maybe [IntegerParameterRange])

The array of IntegerParameterRange objects that specify ranges of integer hyperparameters that a hyperparameter tuning job searches.

parameterRanges_continuousParameterRanges :: Lens' ParameterRanges (Maybe [ContinuousParameterRange])

The array of ContinuousParameterRange objects that specify ranges of continuous hyperparameters that a hyperparameter tuning job searches.