Amazonka.SageMaker.Types.PipelineExecutionStep

Description

Documentation

data PipelineExecutionStep

An execution of a step in a pipeline.

See: newPipelineExecutionStep smart constructor.

Constructors

PipelineExecutionStep'

Fields

failureReason :: Maybe Text

The reason why the step failed execution. This is only returned if the step failed its execution.

startTime :: Maybe POSIX

The time that the step started executing.

stepName :: Maybe Text

The name of the step that is executed.

stepStatus :: Maybe StepStatus

The status of the step execution.

endTime :: Maybe POSIX

The time that the step stopped executing.

metadata :: Maybe PipelineExecutionStepMetadata

Metadata for the step execution.

cacheHitResult :: Maybe CacheHitResult

If this pipeline execution step was cached, details on the cache hit.

newPipelineExecutionStep :: PipelineExecutionStep

Create a value of PipelineExecutionStep with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:PipelineExecutionStep', pipelineExecutionStep_failureReason - The reason why the step failed execution. This is only returned if the step failed its execution.

$sel:startTime:PipelineExecutionStep', pipelineExecutionStep_startTime - The time that the step started executing.

$sel:stepName:PipelineExecutionStep', pipelineExecutionStep_stepName - The name of the step that is executed.

$sel:stepStatus:PipelineExecutionStep', pipelineExecutionStep_stepStatus - The status of the step execution.

$sel:endTime:PipelineExecutionStep', pipelineExecutionStep_endTime - The time that the step stopped executing.

$sel:metadata:PipelineExecutionStep', pipelineExecutionStep_metadata - Metadata for the step execution.

$sel:cacheHitResult:PipelineExecutionStep', pipelineExecutionStep_cacheHitResult - If this pipeline execution step was cached, details on the cache hit.

pipelineExecutionStep_failureReason :: Lens' PipelineExecutionStep (Maybe Text)

The reason why the step failed execution. This is only returned if the step failed its execution.

pipelineExecutionStep_startTime :: Lens' PipelineExecutionStep (Maybe UTCTime)

The time that the step started executing.

pipelineExecutionStep_stepName :: Lens' PipelineExecutionStep (Maybe Text)

The name of the step that is executed.

pipelineExecutionStep_stepStatus :: Lens' PipelineExecutionStep (Maybe StepStatus)

The status of the step execution.

pipelineExecutionStep_endTime :: Lens' PipelineExecutionStep (Maybe UTCTime)

The time that the step stopped executing.

pipelineExecutionStep_metadata :: Lens' PipelineExecutionStep (Maybe PipelineExecutionStepMetadata)

Metadata for the step execution.

pipelineExecutionStep_cacheHitResult :: Lens' PipelineExecutionStep (Maybe CacheHitResult)

If this pipeline execution step was cached, details on the cache hit.