Amazonka.SageMaker.Types.ProcessingS3Input

Description

Documentation

data ProcessingS3Input

Configuration for downloading input data from Amazon S3 into the processing container.

See: newProcessingS3Input smart constructor.

Constructors

ProcessingS3Input'

Fields

s3DataDistributionType :: Maybe ProcessingS3DataDistributionType

Whether to distribute the data from Amazon S3 to all processing instances with FullyReplicated, or whether the data from Amazon S3 is shared by Amazon S3 key, downloading one shard of data to each processing instance.

s3InputMode :: Maybe ProcessingS3InputMode

Whether to use File or Pipe input mode. In File mode, Amazon SageMaker copies the data from the input source onto the local ML storage volume before starting your processing container. This is the most commonly used input mode. In Pipe mode, Amazon SageMaker streams input data from the source directly to your processing container into named pipes without using the ML storage volume.

localPath :: Maybe Text

The local path in your container where you want Amazon SageMaker to write input data to. LocalPath is an absolute path to the input data and must begin with /opt/ml/processing/. LocalPath is a required parameter when AppManaged is False (default).

s3CompressionType :: Maybe ProcessingS3CompressionType

Whether to GZIP-decompress the data in Amazon S3 as it is streamed into the processing container. Gzip can only be used when Pipe mode is specified as the S3InputMode. In Pipe mode, Amazon SageMaker streams input data from the source directly to your container without using the EBS volume.

s3Uri :: Text

The URI of the Amazon S3 prefix Amazon SageMaker downloads data required to run a processing job.

s3DataType :: ProcessingS3DataType

Whether you use an S3Prefix or a ManifestFile for the data type. If you choose S3Prefix, S3Uri identifies a key name prefix. Amazon SageMaker uses all objects with the specified key name prefix for the processing job. If you choose ManifestFile, S3Uri identifies an object that is a manifest file containing a list of object keys that you want Amazon SageMaker to use for the processing job.

newProcessingS3Input

Arguments

:: Text
$sel:s3Uri:ProcessingS3Input'

-> ProcessingS3DataType
$sel:s3DataType:ProcessingS3Input'

-> ProcessingS3Input

Create a value of ProcessingS3Input with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3DataDistributionType:ProcessingS3Input', processingS3Input_s3DataDistributionType - Whether to distribute the data from Amazon S3 to all processing instances with FullyReplicated, or whether the data from Amazon S3 is shared by Amazon S3 key, downloading one shard of data to each processing instance.

$sel:s3InputMode:ProcessingS3Input', processingS3Input_s3InputMode - Whether to use File or Pipe input mode. In File mode, Amazon SageMaker copies the data from the input source onto the local ML storage volume before starting your processing container. This is the most commonly used input mode. In Pipe mode, Amazon SageMaker streams input data from the source directly to your processing container into named pipes without using the ML storage volume.

$sel:localPath:ProcessingS3Input', processingS3Input_localPath - The local path in your container where you want Amazon SageMaker to write input data to. LocalPath is an absolute path to the input data and must begin with /opt/ml/processing/. LocalPath is a required parameter when AppManaged is False (default).

$sel:s3CompressionType:ProcessingS3Input', processingS3Input_s3CompressionType - Whether to GZIP-decompress the data in Amazon S3 as it is streamed into the processing container. Gzip can only be used when Pipe mode is specified as the S3InputMode. In Pipe mode, Amazon SageMaker streams input data from the source directly to your container without using the EBS volume.

$sel:s3Uri:ProcessingS3Input', processingS3Input_s3Uri - The URI of the Amazon S3 prefix Amazon SageMaker downloads data required to run a processing job.

$sel:s3DataType:ProcessingS3Input', processingS3Input_s3DataType - Whether you use an S3Prefix or a ManifestFile for the data type. If you choose S3Prefix, S3Uri identifies a key name prefix. Amazon SageMaker uses all objects with the specified key name prefix for the processing job. If you choose ManifestFile, S3Uri identifies an object that is a manifest file containing a list of object keys that you want Amazon SageMaker to use for the processing job.

processingS3Input_s3DataDistributionType :: Lens' ProcessingS3Input (Maybe ProcessingS3DataDistributionType)

Whether to distribute the data from Amazon S3 to all processing instances with FullyReplicated, or whether the data from Amazon S3 is shared by Amazon S3 key, downloading one shard of data to each processing instance.

processingS3Input_s3InputMode :: Lens' ProcessingS3Input (Maybe ProcessingS3InputMode)

Whether to use File or Pipe input mode. In File mode, Amazon SageMaker copies the data from the input source onto the local ML storage volume before starting your processing container. This is the most commonly used input mode. In Pipe mode, Amazon SageMaker streams input data from the source directly to your processing container into named pipes without using the ML storage volume.

processingS3Input_localPath :: Lens' ProcessingS3Input (Maybe Text)

The local path in your container where you want Amazon SageMaker to write input data to. LocalPath is an absolute path to the input data and must begin with /opt/ml/processing/. LocalPath is a required parameter when AppManaged is False (default).

processingS3Input_s3CompressionType :: Lens' ProcessingS3Input (Maybe ProcessingS3CompressionType)

Whether to GZIP-decompress the data in Amazon S3 as it is streamed into the processing container. Gzip can only be used when Pipe mode is specified as the S3InputMode. In Pipe mode, Amazon SageMaker streams input data from the source directly to your container without using the EBS volume.

processingS3Input_s3Uri :: Lens' ProcessingS3Input Text

The URI of the Amazon S3 prefix Amazon SageMaker downloads data required to run a processing job.

processingS3Input_s3DataType :: Lens' ProcessingS3Input ProcessingS3DataType

Whether you use an S3Prefix or a ManifestFile for the data type. If you choose S3Prefix, S3Uri identifies a key name prefix. Amazon SageMaker uses all objects with the specified key name prefix for the processing job. If you choose ManifestFile, S3Uri identifies an object that is a manifest file containing a list of object keys that you want Amazon SageMaker to use for the processing job.