Amazonka.SageMaker.Types.ProductionVariantCoreDumpConfig

Description

Documentation

data ProductionVariantCoreDumpConfig

Specifies configuration for a core dump from the model container when the process crashes.

See: newProductionVariantCoreDumpConfig smart constructor.

Constructors

ProductionVariantCoreDumpConfig'

Fields

kmsKeyId :: Maybe Text

The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt the core dump data at rest using Amazon S3 server-side encryption. The KmsKeyId can be any of the following formats:

// KMS Key ID

"1234abcd-12ab-34cd-56ef-1234567890ab"

// Amazon Resource Name (ARN) of a KMS Key

"arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

// KMS Key Alias

"alias/ExampleAlias"

// Amazon Resource Name (ARN) of a KMS Key Alias

"arn:aws:kms:us-west-2:111122223333:alias/ExampleAlias"

If you use a KMS key ID or an alias of your KMS key, the Amazon SageMaker execution role must include permissions to call kms:Encrypt. If you don't provide a KMS key ID, Amazon SageMaker uses the default KMS key for Amazon S3 for your role's account. Amazon SageMaker uses server-side encryption with KMS-managed keys for OutputDataConfig. If you use a bucket policy with an s3:PutObject permission that only allows objects with server-side encryption, set the condition key of s3:x-amz-server-side-encryption to "aws:kms". For more information, see KMS-Managed Encryption Keys in the Amazon Simple Storage Service Developer Guide.

The KMS key policy must grant permission to the IAM role that you specify in your CreateEndpoint and UpdateEndpoint requests. For more information, see Using Key Policies in Amazon Web Services KMS in the Amazon Web Services Key Management Service Developer Guide.

destinationS3Uri :: Text

The Amazon S3 bucket to send the core dump to.

newProductionVariantCoreDumpConfig

Arguments

:: Text
$sel:destinationS3Uri:ProductionVariantCoreDumpConfig'

-> ProductionVariantCoreDumpConfig

Create a value of ProductionVariantCoreDumpConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:ProductionVariantCoreDumpConfig', productionVariantCoreDumpConfig_kmsKeyId - The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt the core dump data at rest using Amazon S3 server-side encryption. The KmsKeyId can be any of the following formats:

// KMS Key ID

"1234abcd-12ab-34cd-56ef-1234567890ab"

// Amazon Resource Name (ARN) of a KMS Key

"arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

// KMS Key Alias

"alias/ExampleAlias"

// Amazon Resource Name (ARN) of a KMS Key Alias

"arn:aws:kms:us-west-2:111122223333:alias/ExampleAlias"

If you use a KMS key ID or an alias of your KMS key, the Amazon SageMaker execution role must include permissions to call kms:Encrypt. If you don't provide a KMS key ID, Amazon SageMaker uses the default KMS key for Amazon S3 for your role's account. Amazon SageMaker uses server-side encryption with KMS-managed keys for OutputDataConfig. If you use a bucket policy with an s3:PutObject permission that only allows objects with server-side encryption, set the condition key of s3:x-amz-server-side-encryption to "aws:kms". For more information, see KMS-Managed Encryption Keys in the Amazon Simple Storage Service Developer Guide.

The KMS key policy must grant permission to the IAM role that you specify in your CreateEndpoint and UpdateEndpoint requests. For more information, see Using Key Policies in Amazon Web Services KMS in the Amazon Web Services Key Management Service Developer Guide.

$sel:destinationS3Uri:ProductionVariantCoreDumpConfig', productionVariantCoreDumpConfig_destinationS3Uri - The Amazon S3 bucket to send the core dump to.

productionVariantCoreDumpConfig_kmsKeyId :: Lens' ProductionVariantCoreDumpConfig (Maybe Text)

The Amazon Web Services Key Management Service (Amazon Web Services KMS) key that Amazon SageMaker uses to encrypt the core dump data at rest using Amazon S3 server-side encryption. The KmsKeyId can be any of the following formats:

// KMS Key ID

"1234abcd-12ab-34cd-56ef-1234567890ab"

// Amazon Resource Name (ARN) of a KMS Key

"arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab"

// KMS Key Alias

"alias/ExampleAlias"

// Amazon Resource Name (ARN) of a KMS Key Alias

"arn:aws:kms:us-west-2:111122223333:alias/ExampleAlias"

If you use a KMS key ID or an alias of your KMS key, the Amazon SageMaker execution role must include permissions to call kms:Encrypt. If you don't provide a KMS key ID, Amazon SageMaker uses the default KMS key for Amazon S3 for your role's account. Amazon SageMaker uses server-side encryption with KMS-managed keys for OutputDataConfig. If you use a bucket policy with an s3:PutObject permission that only allows objects with server-side encryption, set the condition key of s3:x-amz-server-side-encryption to "aws:kms". For more information, see KMS-Managed Encryption Keys in the Amazon Simple Storage Service Developer Guide.

The KMS key policy must grant permission to the IAM role that you specify in your CreateEndpoint and UpdateEndpoint requests. For more information, see Using Key Policies in Amazon Web Services KMS in the Amazon Web Services Key Management Service Developer Guide.

productionVariantCoreDumpConfig_destinationS3Uri :: Lens' ProductionVariantCoreDumpConfig Text

The Amazon S3 bucket to send the core dump to.