Amazonka.SageMaker.Types.ResourceSpec

Description

Documentation

data ResourceSpec

Specifies the ARN's of a SageMaker image and SageMaker image version, and the instance type that the version runs on.

See: newResourceSpec smart constructor.

Constructors

ResourceSpec'

Fields

instanceType :: Maybe AppInstanceType

The instance type that the image version runs on.

sageMakerImageArn :: Maybe Text

The ARN of the SageMaker image that the image version belongs to.

sageMakerImageVersionArn :: Maybe Text

The ARN of the image version created on the instance.

lifecycleConfigArn :: Maybe Text

The Amazon Resource Name (ARN) of the Lifecycle Configuration attached to the Resource.

newResourceSpec :: ResourceSpec

Create a value of ResourceSpec with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceType:ResourceSpec', resourceSpec_instanceType - The instance type that the image version runs on.

$sel:sageMakerImageArn:ResourceSpec', resourceSpec_sageMakerImageArn - The ARN of the SageMaker image that the image version belongs to.

$sel:sageMakerImageVersionArn:ResourceSpec', resourceSpec_sageMakerImageVersionArn - The ARN of the image version created on the instance.

$sel:lifecycleConfigArn:ResourceSpec', resourceSpec_lifecycleConfigArn - The Amazon Resource Name (ARN) of the Lifecycle Configuration attached to the Resource.

resourceSpec_instanceType :: Lens' ResourceSpec (Maybe AppInstanceType)

The instance type that the image version runs on.

resourceSpec_sageMakerImageArn :: Lens' ResourceSpec (Maybe Text)

The ARN of the SageMaker image that the image version belongs to.

resourceSpec_sageMakerImageVersionArn :: Lens' ResourceSpec (Maybe Text)

The ARN of the image version created on the instance.

resourceSpec_lifecycleConfigArn :: Lens' ResourceSpec (Maybe Text)

The Amazon Resource Name (ARN) of the Lifecycle Configuration attached to the Resource.