Amazonka.SageMaker.Types.Workteam

Description

Documentation

data Workteam

Provides details about a labeling work team.

See: newWorkteam smart constructor.

Constructors

Workteam'

Fields

subDomain :: Maybe Text

The URI of the labeling job's user interface. Workers open this URI to start labeling your data objects.

productListingIds :: Maybe [Text]

The Amazon Marketplace identifier for a vendor's work team.

notificationConfiguration :: Maybe NotificationConfiguration

Configures SNS notifications of available or expiring work items for work teams.

createDate :: Maybe POSIX

The date and time that the work team was created (timestamp).

workforceArn :: Maybe Text

The Amazon Resource Name (ARN) of the workforce.

lastUpdatedDate :: Maybe POSIX

The date and time that the work team was last updated (timestamp).

workteamName :: Text

The name of the work team.

memberDefinitions :: NonEmpty MemberDefinition

A list of MemberDefinition objects that contains objects that identify the workers that make up the work team.

Workforces can be created using Amazon Cognito or your own OIDC Identity Provider (IdP). For private workforces created using Amazon Cognito use CognitoMemberDefinition. For workforces created using your own OIDC identity provider (IdP) use OidcMemberDefinition.

workteamArn :: Text

The Amazon Resource Name (ARN) that identifies the work team.

description :: Text

A description of the work team.

newWorkteam

Arguments

:: Text
$sel:workteamName:Workteam'

-> NonEmpty MemberDefinition
$sel:memberDefinitions:Workteam'

-> Text
$sel:workteamArn:Workteam'

-> Text
$sel:description:Workteam'

-> Workteam

Create a value of Workteam with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subDomain:Workteam', workteam_subDomain - The URI of the labeling job's user interface. Workers open this URI to start labeling your data objects.

$sel:productListingIds:Workteam', workteam_productListingIds - The Amazon Marketplace identifier for a vendor's work team.

$sel:notificationConfiguration:Workteam', workteam_notificationConfiguration - Configures SNS notifications of available or expiring work items for work teams.

$sel:createDate:Workteam', workteam_createDate - The date and time that the work team was created (timestamp).

$sel:workforceArn:Workteam', workteam_workforceArn - The Amazon Resource Name (ARN) of the workforce.

$sel:lastUpdatedDate:Workteam', workteam_lastUpdatedDate - The date and time that the work team was last updated (timestamp).

$sel:workteamName:Workteam', workteam_workteamName - The name of the work team.

$sel:memberDefinitions:Workteam', workteam_memberDefinitions - A list of MemberDefinition objects that contains objects that identify the workers that make up the work team.

Workforces can be created using Amazon Cognito or your own OIDC Identity Provider (IdP). For private workforces created using Amazon Cognito use CognitoMemberDefinition. For workforces created using your own OIDC identity provider (IdP) use OidcMemberDefinition.

$sel:workteamArn:Workteam', workteam_workteamArn - The Amazon Resource Name (ARN) that identifies the work team.

$sel:description:Workteam', workteam_description - A description of the work team.

workteam_subDomain :: Lens' Workteam (Maybe Text)

The URI of the labeling job's user interface. Workers open this URI to start labeling your data objects.

workteam_productListingIds :: Lens' Workteam (Maybe [Text])

The Amazon Marketplace identifier for a vendor's work team.

workteam_notificationConfiguration :: Lens' Workteam (Maybe NotificationConfiguration)

Configures SNS notifications of available or expiring work items for work teams.

workteam_createDate :: Lens' Workteam (Maybe UTCTime)

The date and time that the work team was created (timestamp).

workteam_workforceArn :: Lens' Workteam (Maybe Text)

The Amazon Resource Name (ARN) of the workforce.

workteam_lastUpdatedDate :: Lens' Workteam (Maybe UTCTime)

The date and time that the work team was last updated (timestamp).

workteam_workteamName :: Lens' Workteam Text

The name of the work team.

workteam_memberDefinitions :: Lens' Workteam (NonEmpty MemberDefinition)

A list of MemberDefinition objects that contains objects that identify the workers that make up the work team.

Workforces can be created using Amazon Cognito or your own OIDC Identity Provider (IdP). For private workforces created using Amazon Cognito use CognitoMemberDefinition. For workforces created using your own OIDC identity provider (IdP) use OidcMemberDefinition.

workteam_workteamArn :: Lens' Workteam Text

The Amazon Resource Name (ARN) that identifies the work team.

workteam_description :: Lens' Workteam Text

A description of the work team.