Amazonka.SageMaker.UpdatePipeline

Description

Updates a pipeline.

Creating a Request

data UpdatePipeline

See: newUpdatePipeline smart constructor.

Constructors

UpdatePipeline'

Fields

pipelineDisplayName :: Maybe Text

The display name of the pipeline.

pipelineDescription :: Maybe Text

The description of the pipeline.

pipelineDefinition :: Maybe Text

The JSON pipeline definition.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) that the pipeline uses to execute.

pipelineName :: Text

The name of the pipeline to update.

newUpdatePipeline

Arguments

:: Text
$sel:pipelineName:UpdatePipeline'

-> UpdatePipeline

Create a value of UpdatePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineDisplayName:UpdatePipeline', updatePipeline_pipelineDisplayName - The display name of the pipeline.

$sel:pipelineDescription:UpdatePipeline', updatePipeline_pipelineDescription - The description of the pipeline.

$sel:pipelineDefinition:UpdatePipeline', updatePipeline_pipelineDefinition - The JSON pipeline definition.

$sel:roleArn:UpdatePipeline', updatePipeline_roleArn - The Amazon Resource Name (ARN) that the pipeline uses to execute.

$sel:pipelineName:UpdatePipeline', updatePipeline_pipelineName - The name of the pipeline to update.

Request Lenses

updatePipeline_pipelineDisplayName :: Lens' UpdatePipeline (Maybe Text)

The display name of the pipeline.

updatePipeline_pipelineDescription :: Lens' UpdatePipeline (Maybe Text)

The description of the pipeline.

updatePipeline_pipelineDefinition :: Lens' UpdatePipeline (Maybe Text)

The JSON pipeline definition.

updatePipeline_roleArn :: Lens' UpdatePipeline (Maybe Text)

The Amazon Resource Name (ARN) that the pipeline uses to execute.

updatePipeline_pipelineName :: Lens' UpdatePipeline Text

The name of the pipeline to update.

Destructuring the Response

data UpdatePipelineResponse

See: newUpdatePipelineResponse smart constructor.

Constructors

UpdatePipelineResponse'

Fields

pipelineArn :: Maybe Text

The Amazon Resource Name (ARN) of the updated pipeline.

httpStatus :: Int

The response's http status code.

newUpdatePipelineResponse

Arguments

:: Int
$sel:httpStatus:UpdatePipelineResponse'

-> UpdatePipelineResponse

Create a value of UpdatePipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineArn:UpdatePipelineResponse', updatePipelineResponse_pipelineArn - The Amazon Resource Name (ARN) of the updated pipeline.

$sel:httpStatus:UpdatePipelineResponse', updatePipelineResponse_httpStatus - The response's http status code.

Response Lenses

updatePipelineResponse_pipelineArn :: Lens' UpdatePipelineResponse (Maybe Text)

The Amazon Resource Name (ARN) of the updated pipeline.

updatePipelineResponse_httpStatus :: Lens' UpdatePipelineResponse Int

The response's http status code.