Amazonka.SageMaker.UpdateWorkforce

Description

Use this operation to update your workforce. You can use this operation to require that workers use specific IP addresses to work on tasks and to update your OpenID Connect (OIDC) Identity Provider (IdP) workforce configuration.

Use SourceIpConfig to restrict worker access to tasks to a specific range of IP addresses. You specify allowed IP addresses by creating a list of up to ten CIDRs. By default, a workforce isn't restricted to specific IP addresses. If you specify a range of IP addresses, workers who attempt to access tasks using any IP address outside the specified range are denied and get a Not Found error message on the worker portal.

Use OidcConfig to update the configuration of a workforce created using your own OIDC IdP.

You can only update your OIDC IdP configuration when there are no work teams associated with your workforce. You can delete work teams using the operation.

After restricting access to a range of IP addresses or updating your OIDC IdP configuration with this operation, you can view details about your update workforce using the operation.

This operation only applies to private workforces.

Creating a Request

data UpdateWorkforce

See: newUpdateWorkforce smart constructor.

Constructors

UpdateWorkforce'

Fields

sourceIpConfig :: Maybe SourceIpConfig

A list of one to ten worker IP address ranges (CIDRs) that can be used to access tasks assigned to this workforce.

Maximum: Ten CIDR values

oidcConfig :: Maybe OidcConfig

Use this parameter to update your OIDC Identity Provider (IdP) configuration for a workforce made using your own IdP.

workforceName :: Text

The name of the private workforce that you want to update. You can find your workforce name by using the operation.

newUpdateWorkforce

Arguments

:: Text
$sel:workforceName:UpdateWorkforce'

-> UpdateWorkforce

Create a value of UpdateWorkforce with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceIpConfig:UpdateWorkforce', updateWorkforce_sourceIpConfig - A list of one to ten worker IP address ranges (CIDRs) that can be used to access tasks assigned to this workforce.

Maximum: Ten CIDR values

$sel:oidcConfig:UpdateWorkforce', updateWorkforce_oidcConfig - Use this parameter to update your OIDC Identity Provider (IdP) configuration for a workforce made using your own IdP.

$sel:workforceName:UpdateWorkforce', updateWorkforce_workforceName - The name of the private workforce that you want to update. You can find your workforce name by using the operation.

Request Lenses

updateWorkforce_sourceIpConfig :: Lens' UpdateWorkforce (Maybe SourceIpConfig)

A list of one to ten worker IP address ranges (CIDRs) that can be used to access tasks assigned to this workforce.

Maximum: Ten CIDR values

updateWorkforce_oidcConfig :: Lens' UpdateWorkforce (Maybe OidcConfig)

Use this parameter to update your OIDC Identity Provider (IdP) configuration for a workforce made using your own IdP.

updateWorkforce_workforceName :: Lens' UpdateWorkforce Text

The name of the private workforce that you want to update. You can find your workforce name by using the operation.

Destructuring the Response

data UpdateWorkforceResponse

See: newUpdateWorkforceResponse smart constructor.

Constructors

UpdateWorkforceResponse'

Fields

httpStatus :: Int

The response's http status code.

workforce :: Workforce

A single private workforce. You can create one private work force in each Amazon Web Services Region. By default, any workforce-related API operation used in a specific region will apply to the workforce created in that region. To learn how to create a private workforce, see Create a Private Workforce.

newUpdateWorkforceResponse

Arguments

:: Int
$sel:httpStatus:UpdateWorkforceResponse'

-> Workforce
$sel:workforce:UpdateWorkforceResponse'

-> UpdateWorkforceResponse

Create a value of UpdateWorkforceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateWorkforceResponse', updateWorkforceResponse_httpStatus - The response's http status code.

$sel:workforce:UpdateWorkforceResponse', updateWorkforceResponse_workforce - A single private workforce. You can create one private work force in each Amazon Web Services Region. By default, any workforce-related API operation used in a specific region will apply to the workforce created in that region. To learn how to create a private workforce, see Create a Private Workforce.

Response Lenses

updateWorkforceResponse_httpStatus :: Lens' UpdateWorkforceResponse Int

The response's http status code.

updateWorkforceResponse_workforce :: Lens' UpdateWorkforceResponse Workforce

A single private workforce. You can create one private work force in each Amazon Web Services Region. By default, any workforce-related API operation used in a specific region will apply to the workforce created in that region. To learn how to create a private workforce, see Create a Private Workforce.