Amazonka.SageMakerA2IRuntime

Description

Derived from API version 2019-11-07 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Augmented AI (Amazon A2I) adds the benefit of human judgment to any machine learning application. When an AI application can't evaluate data with a high degree of confidence, human reviewers can take over. This human review is called a human review workflow. To create and start a human review workflow, you need three resources: a /worker task template, a flow definition, and a human loop/.

For information about these resources and prerequisites for using Amazon A2I, see Get Started with Amazon Augmented AI in the Amazon SageMaker Developer Guide.

This API reference includes information about API actions and data types that you can use to interact with Amazon A2I programmatically. Use this guide to:

Start a human loop with the StartHumanLoop operation when using Amazon A2I with a custom task type. To learn more about the difference between custom and built-in task types, see Use Task Types . To learn how to start a human loop using this API, see Create and Start a Human Loop for a Custom Task Type in the Amazon SageMaker Developer Guide.

Manage your human loops. You can list all human loops that you have created, describe individual human loops, and stop and delete human loops. To learn more, see Monitor and Manage Your Human Loop in the Amazon SageMaker Developer Guide.

Amazon A2I integrates APIs from various AWS services to create and start human review workflows for those services. To learn how Amazon A2I uses these APIs, see Use APIs in Amazon A2I in the Amazon SageMaker Developer Guide.

Service Configuration

defaultService :: Service

API version 2019-11-07 of the Amazon Augmented AI Runtime SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by SageMakerA2IRuntime.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The request isn't valid. Check the syntax and try again.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Your request has the same name as another active human loop but has different input data. You cannot start two human loops with the same name and different input data.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You exceeded your service quota. Service quotas, also referred to as limits, are the maximum number of service resources or operations for your AWS account. For a list of Amazon A2I service quotes, see Amazon Augmented AI Service Quotes. Delete some resources or request an increase in your service quota. You can request a quota increase using Service Quotas or the AWS Support Center. To request an increase, see AWS Service Quotas in the AWS General Reference.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

You exceeded the maximum number of requests.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

We couldn't process your request because of an issue with the server. Try again later.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

We couldn't find the requested resource. Check that your resources exists and were created in the same AWS Region as your request, and try your request again.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListHumanLoops (Paginated)

data ListHumanLoops

See: newListHumanLoops smart constructor.

Constructors

ListHumanLoops' (Maybe POSIX) (Maybe Text) (Maybe SortOrder) (Maybe POSIX) (Maybe Natural) Text

newListHumanLoops

Arguments

:: Text
$sel:flowDefinitionArn:ListHumanLoops'

-> ListHumanLoops

Create a value of ListHumanLoops with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimeAfter:ListHumanLoops', listHumanLoops_creationTimeAfter - (Optional) The timestamp of the date when you want the human loops to begin in ISO 8601 format. For example, 2020-02-24.

$sel:nextToken:ListHumanLoops', listHumanLoops_nextToken - A token to display the next page of results.

$sel:sortOrder:ListHumanLoops', listHumanLoops_sortOrder - Optional. The order for displaying results. Valid values: Ascending and Descending.

$sel:creationTimeBefore:ListHumanLoops', listHumanLoops_creationTimeBefore - (Optional) The timestamp of the date before which you want the human loops to begin in ISO 8601 format. For example, 2020-02-24.

$sel:maxResults:ListHumanLoops', listHumanLoops_maxResults - The total number of items to return. If the total number of available items is more than the value specified in MaxResults, then a NextToken is returned in the output. You can use this token to display the next page of results.

$sel:flowDefinitionArn:ListHumanLoops', listHumanLoops_flowDefinitionArn - The Amazon Resource Name (ARN) of a flow definition.

data ListHumanLoopsResponse

See: newListHumanLoopsResponse smart constructor.

Constructors

ListHumanLoopsResponse' (Maybe Text) Int [HumanLoopSummary]

newListHumanLoopsResponse

Arguments

:: Int
$sel:httpStatus:ListHumanLoopsResponse'

-> ListHumanLoopsResponse

Create a value of ListHumanLoopsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListHumanLoops', listHumanLoopsResponse_nextToken - A token to display the next page of results.

$sel:httpStatus:ListHumanLoopsResponse', listHumanLoopsResponse_httpStatus - The response's http status code.

$sel:humanLoopSummaries:ListHumanLoopsResponse', listHumanLoopsResponse_humanLoopSummaries - An array of objects that contain information about the human loops.

DeleteHumanLoop

data DeleteHumanLoop

See: newDeleteHumanLoop smart constructor.

Constructors

DeleteHumanLoop' Text

newDeleteHumanLoop

Arguments

:: Text
$sel:humanLoopName:DeleteHumanLoop'

-> DeleteHumanLoop

Create a value of DeleteHumanLoop with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:humanLoopName:DeleteHumanLoop', deleteHumanLoop_humanLoopName - The name of the human loop that you want to delete.

data DeleteHumanLoopResponse

See: newDeleteHumanLoopResponse smart constructor.

Constructors

DeleteHumanLoopResponse' Int

newDeleteHumanLoopResponse

Arguments

:: Int
$sel:httpStatus:DeleteHumanLoopResponse'

-> DeleteHumanLoopResponse

Create a value of DeleteHumanLoopResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteHumanLoopResponse', deleteHumanLoopResponse_httpStatus - The response's http status code.

StartHumanLoop

data StartHumanLoop

See: newStartHumanLoop smart constructor.

Constructors

StartHumanLoop' (Maybe HumanLoopDataAttributes) Text Text HumanLoopInput

newStartHumanLoop

Arguments

:: Text
$sel:humanLoopName:StartHumanLoop'

-> Text
$sel:flowDefinitionArn:StartHumanLoop'

-> HumanLoopInput
$sel:humanLoopInput:StartHumanLoop'

-> StartHumanLoop

Create a value of StartHumanLoop with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataAttributes:StartHumanLoop', startHumanLoop_dataAttributes - Attributes of the specified data. Use DataAttributes to specify if your data is free of personally identifiable information and/or free of adult content.

$sel:humanLoopName:StartHumanLoop', startHumanLoop_humanLoopName - The name of the human loop.

$sel:flowDefinitionArn:StartHumanLoop', startHumanLoop_flowDefinitionArn - The Amazon Resource Name (ARN) of the flow definition associated with this human loop.

$sel:humanLoopInput:StartHumanLoop', startHumanLoop_humanLoopInput - An object that contains information about the human loop.

data StartHumanLoopResponse

See: newStartHumanLoopResponse smart constructor.

Constructors

StartHumanLoopResponse' (Maybe Text) Int

newStartHumanLoopResponse

Arguments

:: Int
$sel:httpStatus:StartHumanLoopResponse'

-> StartHumanLoopResponse

Create a value of StartHumanLoopResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:humanLoopArn:StartHumanLoopResponse', startHumanLoopResponse_humanLoopArn - The Amazon Resource Name (ARN) of the human loop.

$sel:httpStatus:StartHumanLoopResponse', startHumanLoopResponse_httpStatus - The response's http status code.

StopHumanLoop

data StopHumanLoop

See: newStopHumanLoop smart constructor.

Constructors

StopHumanLoop' Text

newStopHumanLoop

Arguments

:: Text
$sel:humanLoopName:StopHumanLoop'

-> StopHumanLoop

Create a value of StopHumanLoop with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:humanLoopName:StopHumanLoop', stopHumanLoop_humanLoopName - The name of the human loop that you want to stop.

data StopHumanLoopResponse

See: newStopHumanLoopResponse smart constructor.

Constructors

StopHumanLoopResponse' Int

newStopHumanLoopResponse

Arguments

:: Int
$sel:httpStatus:StopHumanLoopResponse'

-> StopHumanLoopResponse

Create a value of StopHumanLoopResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StopHumanLoopResponse', stopHumanLoopResponse_httpStatus - The response's http status code.

DescribeHumanLoop

data DescribeHumanLoop

See: newDescribeHumanLoop smart constructor.

Constructors

DescribeHumanLoop' Text

newDescribeHumanLoop

Arguments

:: Text
$sel:humanLoopName:DescribeHumanLoop'

-> DescribeHumanLoop

Create a value of DescribeHumanLoop with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:humanLoopName:DescribeHumanLoop', describeHumanLoop_humanLoopName - The name of the human loop that you want information about.

data DescribeHumanLoopResponse

See: newDescribeHumanLoopResponse smart constructor.

Constructors

DescribeHumanLoopResponse' (Maybe Text) (Maybe Text) (Maybe HumanLoopOutput) Int POSIX HumanLoopStatus Text Text Text

newDescribeHumanLoopResponse

Arguments

:: Int
$sel:httpStatus:DescribeHumanLoopResponse'

-> UTCTime
$sel:creationTime:DescribeHumanLoopResponse'

-> HumanLoopStatus
$sel:humanLoopStatus:DescribeHumanLoopResponse'

-> Text
$sel:humanLoopName:DescribeHumanLoop'

-> Text
$sel:humanLoopArn:DescribeHumanLoopResponse'

-> Text
$sel:flowDefinitionArn:DescribeHumanLoopResponse'

-> DescribeHumanLoopResponse

Create a value of DescribeHumanLoopResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DescribeHumanLoopResponse', describeHumanLoopResponse_failureReason - The reason why a human loop failed. The failure reason is returned when the status of the human loop is Failed.

$sel:failureCode:DescribeHumanLoopResponse', describeHumanLoopResponse_failureCode - A failure code that identifies the type of failure.

Possible values: ValidationError, Expired, InternalError

$sel:humanLoopOutput:DescribeHumanLoopResponse', describeHumanLoopResponse_humanLoopOutput - An object that contains information about the output of the human loop.

$sel:httpStatus:DescribeHumanLoopResponse', describeHumanLoopResponse_httpStatus - The response's http status code.

$sel:creationTime:DescribeHumanLoopResponse', describeHumanLoopResponse_creationTime - The creation time when Amazon Augmented AI created the human loop.

$sel:humanLoopStatus:DescribeHumanLoopResponse', describeHumanLoopResponse_humanLoopStatus - The status of the human loop.

$sel:humanLoopName:DescribeHumanLoop', describeHumanLoopResponse_humanLoopName - The name of the human loop. The name must be lowercase, unique within the Region in your account, and can have up to 63 characters. Valid characters: a-z, 0-9, and - (hyphen).

$sel:humanLoopArn:DescribeHumanLoopResponse', describeHumanLoopResponse_humanLoopArn - The Amazon Resource Name (ARN) of the human loop.

$sel:flowDefinitionArn:DescribeHumanLoopResponse', describeHumanLoopResponse_flowDefinitionArn - The Amazon Resource Name (ARN) of the flow definition.

Types

ContentClassifier

newtype ContentClassifier

Constructors

ContentClassifier'

Fields

fromContentClassifier :: Text

Bundled Patterns

pattern ContentClassifier_FreeOfAdultContent :: ContentClassifier

pattern ContentClassifier_FreeOfPersonallyIdentifiableInformation :: ContentClassifier

HumanLoopStatus

newtype HumanLoopStatus

Constructors

HumanLoopStatus'

Fields

fromHumanLoopStatus :: Text

Bundled Patterns

pattern HumanLoopStatus_Completed :: HumanLoopStatus

pattern HumanLoopStatus_Failed :: HumanLoopStatus

pattern HumanLoopStatus_InProgress :: HumanLoopStatus

pattern HumanLoopStatus_Stopped :: HumanLoopStatus

pattern HumanLoopStatus_Stopping :: HumanLoopStatus

SortOrder

newtype SortOrder

Constructors

SortOrder'

Fields

fromSortOrder :: Text

Bundled Patterns

pattern SortOrder_Ascending :: SortOrder

pattern SortOrder_Descending :: SortOrder

HumanLoopDataAttributes

data HumanLoopDataAttributes

Attributes of the data specified by the customer. Use these to describe the data to be labeled.

See: newHumanLoopDataAttributes smart constructor.

Constructors

HumanLoopDataAttributes' [ContentClassifier]

newHumanLoopDataAttributes :: HumanLoopDataAttributes

Create a value of HumanLoopDataAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentClassifiers:HumanLoopDataAttributes', humanLoopDataAttributes_contentClassifiers - Declares that your content is free of personally identifiable information or adult content.

Amazon SageMaker can restrict the Amazon Mechanical Turk workers who can view your task based on this information.

HumanLoopInput

data HumanLoopInput

An object containing the human loop input in JSON format.

See: newHumanLoopInput smart constructor.

Constructors

HumanLoopInput' Text

newHumanLoopInput

Arguments

:: Text
$sel:inputContent:HumanLoopInput'

-> HumanLoopInput

Create a value of HumanLoopInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputContent:HumanLoopInput', humanLoopInput_inputContent - Serialized input from the human loop. The input must be a string representation of a file in JSON format.

HumanLoopOutput

data HumanLoopOutput

Information about where the human output will be stored.

See: newHumanLoopOutput smart constructor.

Constructors

HumanLoopOutput' Text

newHumanLoopOutput

Arguments

:: Text
$sel:outputS3Uri:HumanLoopOutput'

-> HumanLoopOutput

Create a value of HumanLoopOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputS3Uri:HumanLoopOutput', humanLoopOutput_outputS3Uri - The location of the Amazon S3 object where Amazon Augmented AI stores your human loop output.

HumanLoopSummary

data HumanLoopSummary

Summary information about the human loop.

See: newHumanLoopSummary smart constructor.

Constructors

HumanLoopSummary' (Maybe POSIX) (Maybe Text) (Maybe HumanLoopStatus) (Maybe Text) (Maybe Text)

newHumanLoopSummary :: HumanLoopSummary

Create a value of HumanLoopSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:HumanLoopSummary', humanLoopSummary_creationTime - When Amazon Augmented AI created the human loop.

$sel:failureReason:HumanLoopSummary', humanLoopSummary_failureReason - The reason why the human loop failed. A failure reason is returned when the status of the human loop is Failed.

$sel:humanLoopStatus:HumanLoopSummary', humanLoopSummary_humanLoopStatus - The status of the human loop.

$sel:humanLoopName:HumanLoopSummary', humanLoopSummary_humanLoopName - The name of the human loop.

$sel:flowDefinitionArn:HumanLoopSummary', humanLoopSummary_flowDefinitionArn - The Amazon Resource Name (ARN) of the flow definition used to configure the human loop.