Amazonka.SageMakerA2IRuntime.Lens

Description

Operations

ListHumanLoops

listHumanLoops_creationTimeAfter :: Lens' ListHumanLoops (Maybe UTCTime)

(Optional) The timestamp of the date when you want the human loops to begin in ISO 8601 format. For example, 2020-02-24.

listHumanLoops_nextToken :: Lens' ListHumanLoops (Maybe Text)

A token to display the next page of results.

listHumanLoops_sortOrder :: Lens' ListHumanLoops (Maybe SortOrder)

Optional. The order for displaying results. Valid values: Ascending and Descending.

listHumanLoops_creationTimeBefore :: Lens' ListHumanLoops (Maybe UTCTime)

(Optional) The timestamp of the date before which you want the human loops to begin in ISO 8601 format. For example, 2020-02-24.

listHumanLoops_maxResults :: Lens' ListHumanLoops (Maybe Natural)

The total number of items to return. If the total number of available items is more than the value specified in MaxResults, then a NextToken is returned in the output. You can use this token to display the next page of results.

listHumanLoops_flowDefinitionArn :: Lens' ListHumanLoops Text

The Amazon Resource Name (ARN) of a flow definition.

listHumanLoopsResponse_nextToken :: Lens' ListHumanLoopsResponse (Maybe Text)

A token to display the next page of results.

listHumanLoopsResponse_httpStatus :: Lens' ListHumanLoopsResponse Int

The response's http status code.

listHumanLoopsResponse_humanLoopSummaries :: Lens' ListHumanLoopsResponse [HumanLoopSummary]

An array of objects that contain information about the human loops.

DeleteHumanLoop

deleteHumanLoop_humanLoopName :: Lens' DeleteHumanLoop Text

The name of the human loop that you want to delete.

deleteHumanLoopResponse_httpStatus :: Lens' DeleteHumanLoopResponse Int

The response's http status code.

StartHumanLoop

startHumanLoop_dataAttributes :: Lens' StartHumanLoop (Maybe HumanLoopDataAttributes)

Attributes of the specified data. Use DataAttributes to specify if your data is free of personally identifiable information and/or free of adult content.

startHumanLoop_humanLoopName :: Lens' StartHumanLoop Text

The name of the human loop.

startHumanLoop_flowDefinitionArn :: Lens' StartHumanLoop Text

The Amazon Resource Name (ARN) of the flow definition associated with this human loop.

startHumanLoop_humanLoopInput :: Lens' StartHumanLoop HumanLoopInput

An object that contains information about the human loop.

startHumanLoopResponse_humanLoopArn :: Lens' StartHumanLoopResponse (Maybe Text)

The Amazon Resource Name (ARN) of the human loop.

startHumanLoopResponse_httpStatus :: Lens' StartHumanLoopResponse Int

The response's http status code.

StopHumanLoop

stopHumanLoop_humanLoopName :: Lens' StopHumanLoop Text

The name of the human loop that you want to stop.

stopHumanLoopResponse_httpStatus :: Lens' StopHumanLoopResponse Int

The response's http status code.

DescribeHumanLoop

describeHumanLoop_humanLoopName :: Lens' DescribeHumanLoop Text

The name of the human loop that you want information about.

describeHumanLoopResponse_failureReason :: Lens' DescribeHumanLoopResponse (Maybe Text)

The reason why a human loop failed. The failure reason is returned when the status of the human loop is Failed.

describeHumanLoopResponse_failureCode :: Lens' DescribeHumanLoopResponse (Maybe Text)

A failure code that identifies the type of failure.

Possible values: ValidationError, Expired, InternalError

describeHumanLoopResponse_humanLoopOutput :: Lens' DescribeHumanLoopResponse (Maybe HumanLoopOutput)

An object that contains information about the output of the human loop.

describeHumanLoopResponse_httpStatus :: Lens' DescribeHumanLoopResponse Int

The response's http status code.

describeHumanLoopResponse_creationTime :: Lens' DescribeHumanLoopResponse UTCTime

The creation time when Amazon Augmented AI created the human loop.

describeHumanLoopResponse_humanLoopStatus :: Lens' DescribeHumanLoopResponse HumanLoopStatus

The status of the human loop.

describeHumanLoopResponse_humanLoopName :: Lens' DescribeHumanLoopResponse Text

The name of the human loop. The name must be lowercase, unique within the Region in your account, and can have up to 63 characters. Valid characters: a-z, 0-9, and - (hyphen).

describeHumanLoopResponse_humanLoopArn :: Lens' DescribeHumanLoopResponse Text

The Amazon Resource Name (ARN) of the human loop.

describeHumanLoopResponse_flowDefinitionArn :: Lens' DescribeHumanLoopResponse Text

The Amazon Resource Name (ARN) of the flow definition.

Types

HumanLoopDataAttributes

humanLoopDataAttributes_contentClassifiers :: Lens' HumanLoopDataAttributes [ContentClassifier]

Declares that your content is free of personally identifiable information or adult content.

Amazon SageMaker can restrict the Amazon Mechanical Turk workers who can view your task based on this information.

HumanLoopInput

humanLoopInput_inputContent :: Lens' HumanLoopInput Text

Serialized input from the human loop. The input must be a string representation of a file in JSON format.

HumanLoopOutput

humanLoopOutput_outputS3Uri :: Lens' HumanLoopOutput Text

The location of the Amazon S3 object where Amazon Augmented AI stores your human loop output.

HumanLoopSummary

humanLoopSummary_creationTime :: Lens' HumanLoopSummary (Maybe UTCTime)

When Amazon Augmented AI created the human loop.

humanLoopSummary_failureReason :: Lens' HumanLoopSummary (Maybe Text)

The reason why the human loop failed. A failure reason is returned when the status of the human loop is Failed.

humanLoopSummary_humanLoopStatus :: Lens' HumanLoopSummary (Maybe HumanLoopStatus)

The status of the human loop.

humanLoopSummary_humanLoopName :: Lens' HumanLoopSummary (Maybe Text)

The name of the human loop.

humanLoopSummary_flowDefinitionArn :: Lens' HumanLoopSummary (Maybe Text)

The Amazon Resource Name (ARN) of the flow definition used to configure the human loop.