Amazonka.SageMakerEdge

Description

Derived from API version 2020-09-23 of the AWS service descriptions, licensed under Apache 2.0.

SageMaker Edge Manager dataplane service for communicating with active agents.

Service Configuration

defaultService :: Service

API version 2020-09-23 of the Amazon Sagemaker Edge Manager SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by SageMakerEdge.

InternalServiceException

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

An internal failure occurred. Try your request again. If the problem persists, contact AWS customer support.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

SendHeartbeat

data SendHeartbeat

See: newSendHeartbeat smart constructor.

Constructors

SendHeartbeat' (Maybe [EdgeMetric]) (Maybe [Model]) Text Text Text

newSendHeartbeat

Arguments

:: Text
$sel:agentVersion:SendHeartbeat'

-> Text
$sel:deviceName:SendHeartbeat'

-> Text
$sel:deviceFleetName:SendHeartbeat'

-> SendHeartbeat

Create a value of SendHeartbeat with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentMetrics:SendHeartbeat', sendHeartbeat_agentMetrics - For internal use. Returns a list of SageMaker Edge Manager agent operating metrics.

$sel:models:SendHeartbeat', sendHeartbeat_models - Returns a list of models deployed on the the device.

$sel:agentVersion:SendHeartbeat', sendHeartbeat_agentVersion - Returns the version of the agent.

$sel:deviceName:SendHeartbeat', sendHeartbeat_deviceName - The unique name of the device.

$sel:deviceFleetName:SendHeartbeat', sendHeartbeat_deviceFleetName - The name of the fleet that the device belongs to.

data SendHeartbeatResponse

See: newSendHeartbeatResponse smart constructor.

Constructors

SendHeartbeatResponse'

newSendHeartbeatResponse :: SendHeartbeatResponse

Create a value of SendHeartbeatResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

GetDeviceRegistration

data GetDeviceRegistration

See: newGetDeviceRegistration smart constructor.

Constructors

GetDeviceRegistration' Text Text

newGetDeviceRegistration

Arguments

:: Text
$sel:deviceName:GetDeviceRegistration'

-> Text
$sel:deviceFleetName:GetDeviceRegistration'

-> GetDeviceRegistration

Create a value of GetDeviceRegistration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceName:GetDeviceRegistration', getDeviceRegistration_deviceName - The unique name of the device you want to get the registration status from.

$sel:deviceFleetName:GetDeviceRegistration', getDeviceRegistration_deviceFleetName - The name of the fleet that the device belongs to.

data GetDeviceRegistrationResponse

See: newGetDeviceRegistrationResponse smart constructor.

Constructors

GetDeviceRegistrationResponse' (Maybe Text) (Maybe Text) Int

newGetDeviceRegistrationResponse

Arguments

:: Int
$sel:httpStatus:GetDeviceRegistrationResponse'

-> GetDeviceRegistrationResponse

Create a value of GetDeviceRegistrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cacheTTL:GetDeviceRegistrationResponse', getDeviceRegistrationResponse_cacheTTL - The amount of time, in seconds, that the registration status is stored on the device’s cache before it is refreshed.

$sel:deviceRegistration:GetDeviceRegistrationResponse', getDeviceRegistrationResponse_deviceRegistration - Describes if the device is currently registered with SageMaker Edge Manager.

$sel:httpStatus:GetDeviceRegistrationResponse', getDeviceRegistrationResponse_httpStatus - The response's http status code.

Types

EdgeMetric

data EdgeMetric

Information required for edge device metrics.

See: newEdgeMetric smart constructor.

Constructors

EdgeMetric' (Maybe Text) (Maybe Text) (Maybe Double) (Maybe POSIX)

newEdgeMetric :: EdgeMetric

Create a value of EdgeMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimension:EdgeMetric', edgeMetric_dimension - The dimension of metrics published.

$sel:metricName:EdgeMetric', edgeMetric_metricName - Returns the name of the metric.

$sel:value:EdgeMetric', edgeMetric_value - Returns the value of the metric.

$sel:timestamp:EdgeMetric', edgeMetric_timestamp - Timestamp of when the metric was requested.

Model

data Model

Information about a model deployed on an edge device that is registered with SageMaker Edge Manager.

See: newModel smart constructor.

Constructors

Model' (Maybe Text) (Maybe [EdgeMetric]) (Maybe Text) (Maybe POSIX) (Maybe POSIX)

newModel :: Model

Create a value of Model with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelName:Model', model_modelName - The name of the model.

$sel:modelMetrics:Model', model_modelMetrics - Information required for model metrics.

$sel:modelVersion:Model', model_modelVersion - The version of the model.

$sel:latestInference:Model', model_latestInference - The timestamp of the last inference that was made.

$sel:latestSampleTime:Model', model_latestSampleTime - The timestamp of the last data sample taken.