Amazonka.SageMakerEdge.Lens

Description

Operations

SendHeartbeat

sendHeartbeat_agentMetrics :: Lens' SendHeartbeat (Maybe [EdgeMetric])

For internal use. Returns a list of SageMaker Edge Manager agent operating metrics.

sendHeartbeat_models :: Lens' SendHeartbeat (Maybe [Model])

Returns a list of models deployed on the the device.

sendHeartbeat_agentVersion :: Lens' SendHeartbeat Text

Returns the version of the agent.

sendHeartbeat_deviceName :: Lens' SendHeartbeat Text

The unique name of the device.

sendHeartbeat_deviceFleetName :: Lens' SendHeartbeat Text

The name of the fleet that the device belongs to.

GetDeviceRegistration

getDeviceRegistration_deviceName :: Lens' GetDeviceRegistration Text

The unique name of the device you want to get the registration status from.

getDeviceRegistration_deviceFleetName :: Lens' GetDeviceRegistration Text

The name of the fleet that the device belongs to.

getDeviceRegistrationResponse_cacheTTL :: Lens' GetDeviceRegistrationResponse (Maybe Text)

The amount of time, in seconds, that the registration status is stored on the device’s cache before it is refreshed.

getDeviceRegistrationResponse_deviceRegistration :: Lens' GetDeviceRegistrationResponse (Maybe Text)

Describes if the device is currently registered with SageMaker Edge Manager.

getDeviceRegistrationResponse_httpStatus :: Lens' GetDeviceRegistrationResponse Int

The response's http status code.

Types

EdgeMetric

edgeMetric_dimension :: Lens' EdgeMetric (Maybe Text)

The dimension of metrics published.

edgeMetric_metricName :: Lens' EdgeMetric (Maybe Text)

Returns the name of the metric.

edgeMetric_value :: Lens' EdgeMetric (Maybe Double)

Returns the value of the metric.

edgeMetric_timestamp :: Lens' EdgeMetric (Maybe UTCTime)

Timestamp of when the metric was requested.

Model

model_modelName :: Lens' Model (Maybe Text)

The name of the model.

model_modelMetrics :: Lens' Model (Maybe [EdgeMetric])

Information required for model metrics.

model_modelVersion :: Lens' Model (Maybe Text)

The version of the model.

model_latestInference :: Lens' Model (Maybe UTCTime)

The timestamp of the last inference that was made.

model_latestSampleTime :: Lens' Model (Maybe UTCTime)

The timestamp of the last data sample taken.