Amazonka.SageMakerEdge.SendHeartbeat

Description

Use to get the current status of devices registered on SageMaker Edge Manager.

Creating a Request

data SendHeartbeat

See: newSendHeartbeat smart constructor.

Constructors

SendHeartbeat'

Fields

agentMetrics :: Maybe [EdgeMetric]

For internal use. Returns a list of SageMaker Edge Manager agent operating metrics.

models :: Maybe [Model]

Returns a list of models deployed on the the device.

agentVersion :: Text

Returns the version of the agent.

deviceName :: Text

The unique name of the device.

deviceFleetName :: Text

The name of the fleet that the device belongs to.

newSendHeartbeat

Arguments

:: Text
$sel:agentVersion:SendHeartbeat'

-> Text
$sel:deviceName:SendHeartbeat'

-> Text
$sel:deviceFleetName:SendHeartbeat'

-> SendHeartbeat

Create a value of SendHeartbeat with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentMetrics:SendHeartbeat', sendHeartbeat_agentMetrics - For internal use. Returns a list of SageMaker Edge Manager agent operating metrics.

$sel:models:SendHeartbeat', sendHeartbeat_models - Returns a list of models deployed on the the device.

$sel:agentVersion:SendHeartbeat', sendHeartbeat_agentVersion - Returns the version of the agent.

$sel:deviceName:SendHeartbeat', sendHeartbeat_deviceName - The unique name of the device.

$sel:deviceFleetName:SendHeartbeat', sendHeartbeat_deviceFleetName - The name of the fleet that the device belongs to.

Request Lenses

sendHeartbeat_agentMetrics :: Lens' SendHeartbeat (Maybe [EdgeMetric])

For internal use. Returns a list of SageMaker Edge Manager agent operating metrics.

sendHeartbeat_models :: Lens' SendHeartbeat (Maybe [Model])

Returns a list of models deployed on the the device.

sendHeartbeat_agentVersion :: Lens' SendHeartbeat Text

Returns the version of the agent.

sendHeartbeat_deviceName :: Lens' SendHeartbeat Text

The unique name of the device.

sendHeartbeat_deviceFleetName :: Lens' SendHeartbeat Text

The name of the fleet that the device belongs to.

Destructuring the Response

data SendHeartbeatResponse

See: newSendHeartbeatResponse smart constructor.

Constructors

SendHeartbeatResponse'

newSendHeartbeatResponse :: SendHeartbeatResponse

Create a value of SendHeartbeatResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.