Amazonka.SageMakerEdge.Types.Model

Description

Documentation

data Model

Information about a model deployed on an edge device that is registered with SageMaker Edge Manager.

See: newModel smart constructor.

Constructors

Model'

Fields

modelName :: Maybe Text

The name of the model.

modelMetrics :: Maybe [EdgeMetric]

Information required for model metrics.

modelVersion :: Maybe Text

The version of the model.

latestInference :: Maybe POSIX

The timestamp of the last inference that was made.

latestSampleTime :: Maybe POSIX

The timestamp of the last data sample taken.

newModel :: Model

Create a value of Model with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelName:Model', model_modelName - The name of the model.

$sel:modelMetrics:Model', model_modelMetrics - Information required for model metrics.

$sel:modelVersion:Model', model_modelVersion - The version of the model.

$sel:latestInference:Model', model_latestInference - The timestamp of the last inference that was made.

$sel:latestSampleTime:Model', model_latestSampleTime - The timestamp of the last data sample taken.

model_modelName :: Lens' Model (Maybe Text)

The name of the model.

model_modelMetrics :: Lens' Model (Maybe [EdgeMetric])

Information required for model metrics.

model_modelVersion :: Lens' Model (Maybe Text)

The version of the model.

model_latestInference :: Lens' Model (Maybe UTCTime)

The timestamp of the last inference that was made.

model_latestSampleTime :: Lens' Model (Maybe UTCTime)

The timestamp of the last data sample taken.