Amazonka.SageMakerFeatureStoreRuntime

Description

Derived from API version 2020-07-01 of the AWS service descriptions, licensed under Apache 2.0.

Contains all data plane API operations and data types for the Amazon SageMaker Feature Store. Use this API to put, delete, and retrieve (get) features from a feature store.

Use the following operations to configure your OnlineStore and OfflineStore features, and to create and manage feature groups:

CreateFeatureGroup

DeleteFeatureGroup

DescribeFeatureGroup

ListFeatureGroups

Service Configuration

defaultService :: Service

API version 2020-07-01 of the Amazon SageMaker Feature Store Runtime SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by SageMakerFeatureStoreRuntime.

AccessForbidden

_AccessForbidden :: AsError a => Getting (First ServiceError) a ServiceError

You do not have permission to perform an action.

ServiceUnavailable

_ServiceUnavailable :: AsError a => Getting (First ServiceError) a ServiceError

The service is currently unavailable.

InternalFailure

_InternalFailure :: AsError a => Getting (First ServiceError) a ServiceError

An internal failure occurred. Try your request again. If the problem persists, contact AWS customer support.

ValidationError

_ValidationError :: AsError a => Getting (First ServiceError) a ServiceError

There was an error validating your request.

ResourceNotFound

_ResourceNotFound :: AsError a => Getting (First ServiceError) a ServiceError

A resource that is required to perform an action was not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

PutRecord

data PutRecord

See: newPutRecord smart constructor.

Constructors

PutRecord' Text (NonEmpty FeatureValue)

newPutRecord

Arguments

:: Text
$sel:featureGroupName:PutRecord'

-> NonEmpty FeatureValue
$sel:record:PutRecord'

-> PutRecord

Create a value of PutRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureGroupName:PutRecord', putRecord_featureGroupName - The name of the feature group that you want to insert the record into.

$sel:record:PutRecord', putRecord_record - List of FeatureValues to be inserted. This will be a full over-write. If you only want to update few of the feature values, do the following:

Use GetRecord to retrieve the latest record.

Update the record returned from GetRecord.

Use PutRecord to update feature values.

data PutRecordResponse

See: newPutRecordResponse smart constructor.

Constructors

PutRecordResponse'

newPutRecordResponse :: PutRecordResponse

Create a value of PutRecordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

DeleteRecord

data DeleteRecord

See: newDeleteRecord smart constructor.

Constructors

DeleteRecord' Text Text Text

newDeleteRecord

Arguments

:: Text
$sel:featureGroupName:DeleteRecord'

-> Text
$sel:recordIdentifierValueAsString:DeleteRecord'

-> Text
$sel:eventTime:DeleteRecord'

-> DeleteRecord

Create a value of DeleteRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureGroupName:DeleteRecord', deleteRecord_featureGroupName - The name of the feature group to delete the record from.

$sel:recordIdentifierValueAsString:DeleteRecord', deleteRecord_recordIdentifierValueAsString - The value for the RecordIdentifier that uniquely identifies the record, in string format.

$sel:eventTime:DeleteRecord', deleteRecord_eventTime - Timestamp indicating when the deletion event occurred. EventTime can be used to query data at a certain point in time.

data DeleteRecordResponse

See: newDeleteRecordResponse smart constructor.

Constructors

DeleteRecordResponse'

newDeleteRecordResponse :: DeleteRecordResponse

Create a value of DeleteRecordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

BatchGetRecord

data BatchGetRecord

See: newBatchGetRecord smart constructor.

Constructors

BatchGetRecord' (NonEmpty BatchGetRecordIdentifier)

newBatchGetRecord

Arguments

:: NonEmpty BatchGetRecordIdentifier
$sel:identifiers:BatchGetRecord'

-> BatchGetRecord

Create a value of BatchGetRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identifiers:BatchGetRecord', batchGetRecord_identifiers - A list of FeatureGroup names, with their corresponding RecordIdentifier value, and Feature name that have been requested to be retrieved in batch.

data BatchGetRecordResponse

See: newBatchGetRecordResponse smart constructor.

Constructors

BatchGetRecordResponse' Int [BatchGetRecordResultDetail] [BatchGetRecordError] [BatchGetRecordIdentifier]

newBatchGetRecordResponse

Arguments

:: Int
$sel:httpStatus:BatchGetRecordResponse'

-> BatchGetRecordResponse

Create a value of BatchGetRecordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:BatchGetRecordResponse', batchGetRecordResponse_httpStatus - The response's http status code.

$sel:records:BatchGetRecordResponse', batchGetRecordResponse_records - A list of Records you requested to be retrieved in batch.

$sel:errors:BatchGetRecordResponse', batchGetRecordResponse_errors - A list of errors that have occured when retrieving a batch of Records.

$sel:unprocessedIdentifiers:BatchGetRecordResponse', batchGetRecordResponse_unprocessedIdentifiers - A unprocessed list of FeatureGroup names, with their corresponding RecordIdentifier value, and Feature name.

GetRecord

data GetRecord

See: newGetRecord smart constructor.

Constructors

GetRecord' (Maybe (NonEmpty Text)) Text Text

newGetRecord

Arguments

:: Text
$sel:featureGroupName:GetRecord'

-> Text
$sel:recordIdentifierValueAsString:GetRecord'

-> GetRecord

Create a value of GetRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureNames:GetRecord', getRecord_featureNames - List of names of Features to be retrieved. If not specified, the latest value for all the Features are returned.

$sel:featureGroupName:GetRecord', getRecord_featureGroupName - The name of the feature group in which you want to put the records.

$sel:recordIdentifierValueAsString:GetRecord', getRecord_recordIdentifierValueAsString - The value that corresponds to RecordIdentifier type and uniquely identifies the record in the FeatureGroup.

data GetRecordResponse

See: newGetRecordResponse smart constructor.

Constructors

GetRecordResponse' (Maybe (NonEmpty FeatureValue)) Int

newGetRecordResponse

Arguments

:: Int
$sel:httpStatus:GetRecordResponse'

-> GetRecordResponse

Create a value of GetRecordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:record:GetRecordResponse', getRecordResponse_record - The record you requested. A list of FeatureValues.

$sel:httpStatus:GetRecordResponse', getRecordResponse_httpStatus - The response's http status code.

Types

BatchGetRecordError

data BatchGetRecordError

The error that has occurred when attempting to retrieve a batch of Records.

See: newBatchGetRecordError smart constructor.

Constructors

BatchGetRecordError' Text Text Text Text

newBatchGetRecordError

Arguments

:: Text
$sel:featureGroupName:BatchGetRecordError'

-> Text
$sel:recordIdentifierValueAsString:BatchGetRecordError'

-> Text
$sel:errorCode:BatchGetRecordError'

-> Text
$sel:errorMessage:BatchGetRecordError'

-> BatchGetRecordError

Create a value of BatchGetRecordError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureGroupName:BatchGetRecordError', batchGetRecordError_featureGroupName - The name of the feature group that the record belongs to.

$sel:recordIdentifierValueAsString:BatchGetRecordError', batchGetRecordError_recordIdentifierValueAsString - The value for the RecordIdentifier in string format of a Record from a FeatureGroup that is causing an error when attempting to be retrieved.

$sel:errorCode:BatchGetRecordError', batchGetRecordError_errorCode - The error code of an error that has occured when attempting to retrieve a batch of Records. For more information on errors, see Errors.

$sel:errorMessage:BatchGetRecordError', batchGetRecordError_errorMessage - The error message of an error that has occured when attempting to retrieve a record in the batch.

BatchGetRecordIdentifier

data BatchGetRecordIdentifier

The identifier that identifies the batch of Records you are retrieving in a batch.

See: newBatchGetRecordIdentifier smart constructor.

Constructors

BatchGetRecordIdentifier' (Maybe (NonEmpty Text)) Text (NonEmpty Text)

newBatchGetRecordIdentifier

Arguments

:: Text
$sel:featureGroupName:BatchGetRecordIdentifier'

-> NonEmpty Text
$sel:recordIdentifiersValueAsString:BatchGetRecordIdentifier'

-> BatchGetRecordIdentifier

Create a value of BatchGetRecordIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureNames:BatchGetRecordIdentifier', batchGetRecordIdentifier_featureNames - List of names of Features to be retrieved. If not specified, the latest value for all the Features are returned.

$sel:featureGroupName:BatchGetRecordIdentifier', batchGetRecordIdentifier_featureGroupName - A FeatureGroupName containing Records you are retrieving in a batch.

$sel:recordIdentifiersValueAsString:BatchGetRecordIdentifier', batchGetRecordIdentifier_recordIdentifiersValueAsString - The value for a list of record identifiers in string format.

BatchGetRecordResultDetail

data BatchGetRecordResultDetail

The output of Records that have been retrieved in a batch.

See: newBatchGetRecordResultDetail smart constructor.

Constructors

BatchGetRecordResultDetail' Text Text (NonEmpty FeatureValue)

newBatchGetRecordResultDetail

Arguments

:: Text
$sel:featureGroupName:BatchGetRecordResultDetail'

-> Text
$sel:recordIdentifierValueAsString:BatchGetRecordResultDetail'

-> NonEmpty FeatureValue
$sel:record:BatchGetRecordResultDetail'

-> BatchGetRecordResultDetail

Create a value of BatchGetRecordResultDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureGroupName:BatchGetRecordResultDetail', batchGetRecordResultDetail_featureGroupName - The FeatureGroupName containing Records you retrieved in a batch.

$sel:recordIdentifierValueAsString:BatchGetRecordResultDetail', batchGetRecordResultDetail_recordIdentifierValueAsString - The value of the record identifer in string format.

$sel:record:BatchGetRecordResultDetail', batchGetRecordResultDetail_record - The Record retrieved.

FeatureValue

data FeatureValue

The value associated with a feature.

See: newFeatureValue smart constructor.

Constructors

FeatureValue' Text Text

newFeatureValue

Arguments

:: Text
$sel:featureName:FeatureValue'

-> Text
$sel:valueAsString:FeatureValue'

-> FeatureValue

Create a value of FeatureValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureName:FeatureValue', featureValue_featureName - The name of a feature that a feature value corresponds to.

$sel:valueAsString:FeatureValue', featureValue_valueAsString - The value associated with a feature, in string format. Note that features types can be String, Integral, or Fractional. This value represents all three types as a string.