Amazonka.SageMakerFeatureStoreRuntime.BatchGetRecord

Description

Retrieves a batch of Records from a FeatureGroup.

Creating a Request

data BatchGetRecord

See: newBatchGetRecord smart constructor.

Constructors

BatchGetRecord'

Fields

identifiers :: NonEmpty BatchGetRecordIdentifier

A list of FeatureGroup names, with their corresponding RecordIdentifier value, and Feature name that have been requested to be retrieved in batch.

newBatchGetRecord

Arguments

:: NonEmpty BatchGetRecordIdentifier
$sel:identifiers:BatchGetRecord'

-> BatchGetRecord

Create a value of BatchGetRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identifiers:BatchGetRecord', batchGetRecord_identifiers - A list of FeatureGroup names, with their corresponding RecordIdentifier value, and Feature name that have been requested to be retrieved in batch.

Request Lenses

batchGetRecord_identifiers :: Lens' BatchGetRecord (NonEmpty BatchGetRecordIdentifier)

A list of FeatureGroup names, with their corresponding RecordIdentifier value, and Feature name that have been requested to be retrieved in batch.

Destructuring the Response

data BatchGetRecordResponse

See: newBatchGetRecordResponse smart constructor.

Constructors

BatchGetRecordResponse'

Fields

httpStatus :: Int

The response's http status code.

records :: [BatchGetRecordResultDetail]

A list of Records you requested to be retrieved in batch.

errors :: [BatchGetRecordError]

A list of errors that have occured when retrieving a batch of Records.

unprocessedIdentifiers :: [BatchGetRecordIdentifier]

A unprocessed list of FeatureGroup names, with their corresponding RecordIdentifier value, and Feature name.

newBatchGetRecordResponse

Arguments

:: Int
$sel:httpStatus:BatchGetRecordResponse'

-> BatchGetRecordResponse

Create a value of BatchGetRecordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:BatchGetRecordResponse', batchGetRecordResponse_httpStatus - The response's http status code.

$sel:records:BatchGetRecordResponse', batchGetRecordResponse_records - A list of Records you requested to be retrieved in batch.

$sel:errors:BatchGetRecordResponse', batchGetRecordResponse_errors - A list of errors that have occured when retrieving a batch of Records.

$sel:unprocessedIdentifiers:BatchGetRecordResponse', batchGetRecordResponse_unprocessedIdentifiers - A unprocessed list of FeatureGroup names, with their corresponding RecordIdentifier value, and Feature name.

Response Lenses

batchGetRecordResponse_httpStatus :: Lens' BatchGetRecordResponse Int

The response's http status code.

batchGetRecordResponse_records :: Lens' BatchGetRecordResponse [BatchGetRecordResultDetail]

A list of Records you requested to be retrieved in batch.

batchGetRecordResponse_errors :: Lens' BatchGetRecordResponse [BatchGetRecordError]

A list of errors that have occured when retrieving a batch of Records.

batchGetRecordResponse_unprocessedIdentifiers :: Lens' BatchGetRecordResponse [BatchGetRecordIdentifier]

A unprocessed list of FeatureGroup names, with their corresponding RecordIdentifier value, and Feature name.