Amazonka.SageMakerFeatureStoreRuntime.Types.BatchGetRecordError

Description

Documentation

data BatchGetRecordError

The error that has occurred when attempting to retrieve a batch of Records.

See: newBatchGetRecordError smart constructor.

Constructors

BatchGetRecordError'

Fields

featureGroupName :: Text

The name of the feature group that the record belongs to.

recordIdentifierValueAsString :: Text

The value for the RecordIdentifier in string format of a Record from a FeatureGroup that is causing an error when attempting to be retrieved.

errorCode :: Text

The error code of an error that has occured when attempting to retrieve a batch of Records. For more information on errors, see Errors.

errorMessage :: Text

The error message of an error that has occured when attempting to retrieve a record in the batch.

newBatchGetRecordError

Arguments

:: Text
$sel:featureGroupName:BatchGetRecordError'

-> Text
$sel:recordIdentifierValueAsString:BatchGetRecordError'

-> Text
$sel:errorCode:BatchGetRecordError'

-> Text
$sel:errorMessage:BatchGetRecordError'

-> BatchGetRecordError

Create a value of BatchGetRecordError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureGroupName:BatchGetRecordError', batchGetRecordError_featureGroupName - The name of the feature group that the record belongs to.

$sel:recordIdentifierValueAsString:BatchGetRecordError', batchGetRecordError_recordIdentifierValueAsString - The value for the RecordIdentifier in string format of a Record from a FeatureGroup that is causing an error when attempting to be retrieved.

$sel:errorCode:BatchGetRecordError', batchGetRecordError_errorCode - The error code of an error that has occured when attempting to retrieve a batch of Records. For more information on errors, see Errors.

$sel:errorMessage:BatchGetRecordError', batchGetRecordError_errorMessage - The error message of an error that has occured when attempting to retrieve a record in the batch.

batchGetRecordError_featureGroupName :: Lens' BatchGetRecordError Text

The name of the feature group that the record belongs to.

batchGetRecordError_recordIdentifierValueAsString :: Lens' BatchGetRecordError Text

The value for the RecordIdentifier in string format of a Record from a FeatureGroup that is causing an error when attempting to be retrieved.

batchGetRecordError_errorCode :: Lens' BatchGetRecordError Text

The error code of an error that has occured when attempting to retrieve a batch of Records. For more information on errors, see Errors.

batchGetRecordError_errorMessage :: Lens' BatchGetRecordError Text

The error message of an error that has occured when attempting to retrieve a record in the batch.