Amazonka.SageMakerFeatureStoreRuntime.Types.FeatureValue

Description

Documentation

data FeatureValue

The value associated with a feature.

See: newFeatureValue smart constructor.

Constructors

FeatureValue'

Fields

featureName :: Text

The name of a feature that a feature value corresponds to.

valueAsString :: Text

The value associated with a feature, in string format. Note that features types can be String, Integral, or Fractional. This value represents all three types as a string.

newFeatureValue

Arguments

:: Text
$sel:featureName:FeatureValue'

-> Text
$sel:valueAsString:FeatureValue'

-> FeatureValue

Create a value of FeatureValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureName:FeatureValue', featureValue_featureName - The name of a feature that a feature value corresponds to.

$sel:valueAsString:FeatureValue', featureValue_valueAsString - The value associated with a feature, in string format. Note that features types can be String, Integral, or Fractional. This value represents all three types as a string.

featureValue_featureName :: Lens' FeatureValue Text

The name of a feature that a feature value corresponds to.

featureValue_valueAsString :: Lens' FeatureValue Text

The value associated with a feature, in string format. Note that features types can be String, Integral, or Fractional. This value represents all three types as a string.