Amazonka.SageMakerRuntime

Description

Derived from API version 2017-05-13 of the AWS service descriptions, licensed under Apache 2.0.

The Amazon SageMaker runtime API.

Service Configuration

defaultService :: Service

API version 2017-05-13 of the Amazon SageMaker Runtime SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by SageMakerRuntime.

ServiceUnavailable

_ServiceUnavailable :: AsError a => Getting (First ServiceError) a ServiceError

The service is unavailable. Try your call again.

ModelError

_ModelError :: AsError a => Getting (First ServiceError) a ServiceError

Model (owned by the customer in the container) returned 4xx or 5xx error code.

InternalFailure

_InternalFailure :: AsError a => Getting (First ServiceError) a ServiceError

An internal failure occurred.

ValidationError

_ValidationError :: AsError a => Getting (First ServiceError) a ServiceError

Inspect your request and try again.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

InvokeEndpointAsync

data InvokeEndpointAsync

See: newInvokeEndpointAsync smart constructor.

Constructors

InvokeEndpointAsync' (Maybe Text) (Maybe (Sensitive Text)) (Maybe Text) (Maybe Natural) (Maybe Text) Text Text

newInvokeEndpointAsync

Arguments

:: Text
$sel:endpointName:InvokeEndpointAsync'

-> Text
$sel:inputLocation:InvokeEndpointAsync'

-> InvokeEndpointAsync

Create a value of InvokeEndpointAsync with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accept:InvokeEndpointAsync', invokeEndpointAsync_accept - The desired MIME type of the inference in the response.

$sel:customAttributes:InvokeEndpointAsync', invokeEndpointAsync_customAttributes - Provides additional information about a request for an inference submitted to a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to provide an ID that you can use to track a request or to provide other metadata that a service endpoint was programmed to process. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1).

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

$sel:inferenceId:InvokeEndpointAsync', invokeEndpointAsync_inferenceId - The identifier for the inference request. Amazon SageMaker will generate an identifier for you if none is specified.

$sel:requestTTLSeconds:InvokeEndpointAsync', invokeEndpointAsync_requestTTLSeconds - Maximum age in seconds a request can be in the queue before it is marked as expired.

$sel:contentType:InvokeEndpointAsync', invokeEndpointAsync_contentType - The MIME type of the input data in the request body.

$sel:endpointName:InvokeEndpointAsync', invokeEndpointAsync_endpointName - The name of the endpoint that you specified when you created the endpoint using the CreateEndpoint API.

$sel:inputLocation:InvokeEndpointAsync', invokeEndpointAsync_inputLocation - The Amazon S3 URI where the inference request payload is stored.

data InvokeEndpointAsyncResponse

See: newInvokeEndpointAsyncResponse smart constructor.

Constructors

InvokeEndpointAsyncResponse' (Maybe Text) (Maybe Text) Int

newInvokeEndpointAsyncResponse

Arguments

:: Int
$sel:httpStatus:InvokeEndpointAsyncResponse'

-> InvokeEndpointAsyncResponse

Create a value of InvokeEndpointAsyncResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputLocation:InvokeEndpointAsyncResponse', invokeEndpointAsyncResponse_outputLocation - The Amazon S3 URI where the inference response payload is stored.

$sel:inferenceId:InvokeEndpointAsync', invokeEndpointAsyncResponse_inferenceId - Identifier for an inference request. This will be the same as the InferenceId specified in the input. Amazon SageMaker will generate an identifier for you if you do not specify one.

$sel:httpStatus:InvokeEndpointAsyncResponse', invokeEndpointAsyncResponse_httpStatus - The response's http status code.

InvokeEndpoint

data InvokeEndpoint

See: newInvokeEndpoint smart constructor.

Constructors

InvokeEndpoint' (Maybe Text) (Maybe Text) (Maybe (Sensitive Text)) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Text Text

newInvokeEndpoint

Arguments

:: Text
$sel:endpointName:InvokeEndpoint'

-> Text
$sel:body:InvokeEndpoint'

-> InvokeEndpoint

Create a value of InvokeEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accept:InvokeEndpoint', invokeEndpoint_accept - The desired MIME type of the inference in the response.

$sel:targetModel:InvokeEndpoint', invokeEndpoint_targetModel - The model to request for inference when invoking a multi-model endpoint.

$sel:customAttributes:InvokeEndpoint', invokeEndpoint_customAttributes - Provides additional information about a request for an inference submitted to a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to provide an ID that you can use to track a request or to provide other metadata that a service endpoint was programmed to process. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1).

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

$sel:inferenceId:InvokeEndpoint', invokeEndpoint_inferenceId - If you provide a value, it is added to the captured data when you enable data capture on the endpoint. For information about data capture, see Capture Data.

$sel:targetVariant:InvokeEndpoint', invokeEndpoint_targetVariant - Specify the production variant to send the inference request to when invoking an endpoint that is running two or more variants. Note that this parameter overrides the default behavior for the endpoint, which is to distribute the invocation traffic based on the variant weights.

For information about how to use variant targeting to perform a/b testing, see Test models in production

$sel:contentType:InvokeEndpoint', invokeEndpoint_contentType - The MIME type of the input data in the request body.

$sel:targetContainerHostname:InvokeEndpoint', invokeEndpoint_targetContainerHostname - If the endpoint hosts multiple containers and is configured to use direct invocation, this parameter specifies the host name of the container to invoke.

$sel:endpointName:InvokeEndpoint', invokeEndpoint_endpointName - The name of the endpoint that you specified when you created the endpoint using the CreateEndpoint API.

$sel:body:InvokeEndpoint', invokeEndpoint_body - Provides input data, in the format specified in the ContentType request header. Amazon SageMaker passes all of the data in the body to the model.

For information about the format of the request body, see Common Data Formats-Inference.

data InvokeEndpointResponse

See: newInvokeEndpointResponse smart constructor.

Constructors

InvokeEndpointResponse' (Maybe Text) (Maybe (Sensitive Text)) (Maybe Text) Int Text

newInvokeEndpointResponse

Arguments

:: Int
$sel:httpStatus:InvokeEndpointResponse'

-> Text
$sel:body:InvokeEndpoint'

-> InvokeEndpointResponse

Create a value of InvokeEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invokedProductionVariant:InvokeEndpointResponse', invokeEndpointResponse_invokedProductionVariant - Identifies the production variant that was invoked.

$sel:customAttributes:InvokeEndpoint', invokeEndpointResponse_customAttributes - Provides additional information in the response about the inference returned by a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to return an ID received in the CustomAttributes header of a request or other metadata that a service endpoint was programmed to produce. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1). If the customer wants the custom attribute returned, the model must set the custom attribute to be included on the way back.

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

$sel:contentType:InvokeEndpoint', invokeEndpointResponse_contentType - The MIME type of the inference returned in the response body.

$sel:httpStatus:InvokeEndpointResponse', invokeEndpointResponse_httpStatus - The response's http status code.

$sel:body:InvokeEndpoint', invokeEndpointResponse_body - Includes the inference provided by the model.

For information about the format of the response body, see Common Data Formats-Inference.

Types