Amazonka.SageMakerRuntime.InvokeEndpoint

Description

After you deploy a model into production using Amazon SageMaker hosting services, your client applications use this API to get inferences from the model hosted at the specified endpoint.

For an overview of Amazon SageMaker, see How It Works.

Amazon SageMaker strips all POST headers except those supported by the API. Amazon SageMaker might add additional headers. You should not rely on the behavior of headers outside those enumerated in the request syntax.

Calls to InvokeEndpoint are authenticated by using AWS Signature Version 4. For information, see Authenticating Requests (AWS Signature Version 4) in the Amazon S3 API Reference.

A customer's model containers must respond to requests within 60 seconds. The model itself can have a maximum processing time of 60 seconds before responding to invocations. If your model is going to take 50-60 seconds of processing time, the SDK socket timeout should be set to be 70 seconds.

Endpoints are scoped to an individual account, and are not public. The URL does not contain the account ID, but Amazon SageMaker determines the account ID from the authentication token that is supplied by the caller.

Creating a Request

data InvokeEndpoint

See: newInvokeEndpoint smart constructor.

Constructors

InvokeEndpoint'

Fields

accept :: Maybe Text

The desired MIME type of the inference in the response.

targetModel :: Maybe Text

The model to request for inference when invoking a multi-model endpoint.

customAttributes :: Maybe (Sensitive Text)

Provides additional information about a request for an inference submitted to a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to provide an ID that you can use to track a request or to provide other metadata that a service endpoint was programmed to process. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1).

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

inferenceId :: Maybe Text

If you provide a value, it is added to the captured data when you enable data capture on the endpoint. For information about data capture, see Capture Data.

targetVariant :: Maybe Text

Specify the production variant to send the inference request to when invoking an endpoint that is running two or more variants. Note that this parameter overrides the default behavior for the endpoint, which is to distribute the invocation traffic based on the variant weights.

For information about how to use variant targeting to perform a/b testing, see Test models in production

contentType :: Maybe Text

The MIME type of the input data in the request body.

targetContainerHostname :: Maybe Text

If the endpoint hosts multiple containers and is configured to use direct invocation, this parameter specifies the host name of the container to invoke.

endpointName :: Text

The name of the endpoint that you specified when you created the endpoint using the CreateEndpoint API.

body :: Text

Provides input data, in the format specified in the ContentType request header. Amazon SageMaker passes all of the data in the body to the model.

For information about the format of the request body, see Common Data Formats-Inference.

newInvokeEndpoint

Arguments

:: Text
$sel:endpointName:InvokeEndpoint'

-> Text
$sel:body:InvokeEndpoint'

-> InvokeEndpoint

Create a value of InvokeEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accept:InvokeEndpoint', invokeEndpoint_accept - The desired MIME type of the inference in the response.

$sel:targetModel:InvokeEndpoint', invokeEndpoint_targetModel - The model to request for inference when invoking a multi-model endpoint.

$sel:customAttributes:InvokeEndpoint', invokeEndpoint_customAttributes - Provides additional information about a request for an inference submitted to a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to provide an ID that you can use to track a request or to provide other metadata that a service endpoint was programmed to process. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1).

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

$sel:inferenceId:InvokeEndpoint', invokeEndpoint_inferenceId - If you provide a value, it is added to the captured data when you enable data capture on the endpoint. For information about data capture, see Capture Data.

$sel:targetVariant:InvokeEndpoint', invokeEndpoint_targetVariant - Specify the production variant to send the inference request to when invoking an endpoint that is running two or more variants. Note that this parameter overrides the default behavior for the endpoint, which is to distribute the invocation traffic based on the variant weights.

For information about how to use variant targeting to perform a/b testing, see Test models in production

$sel:contentType:InvokeEndpoint', invokeEndpoint_contentType - The MIME type of the input data in the request body.

$sel:targetContainerHostname:InvokeEndpoint', invokeEndpoint_targetContainerHostname - If the endpoint hosts multiple containers and is configured to use direct invocation, this parameter specifies the host name of the container to invoke.

$sel:endpointName:InvokeEndpoint', invokeEndpoint_endpointName - The name of the endpoint that you specified when you created the endpoint using the CreateEndpoint API.

$sel:body:InvokeEndpoint', invokeEndpoint_body - Provides input data, in the format specified in the ContentType request header. Amazon SageMaker passes all of the data in the body to the model.

For information about the format of the request body, see Common Data Formats-Inference.

Request Lenses

invokeEndpoint_accept :: Lens' InvokeEndpoint (Maybe Text)

The desired MIME type of the inference in the response.

invokeEndpoint_targetModel :: Lens' InvokeEndpoint (Maybe Text)

The model to request for inference when invoking a multi-model endpoint.

invokeEndpoint_customAttributes :: Lens' InvokeEndpoint (Maybe Text)

Provides additional information about a request for an inference submitted to a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to provide an ID that you can use to track a request or to provide other metadata that a service endpoint was programmed to process. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1).

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

invokeEndpoint_inferenceId :: Lens' InvokeEndpoint (Maybe Text)

If you provide a value, it is added to the captured data when you enable data capture on the endpoint. For information about data capture, see Capture Data.

invokeEndpoint_targetVariant :: Lens' InvokeEndpoint (Maybe Text)

Specify the production variant to send the inference request to when invoking an endpoint that is running two or more variants. Note that this parameter overrides the default behavior for the endpoint, which is to distribute the invocation traffic based on the variant weights.

For information about how to use variant targeting to perform a/b testing, see Test models in production

invokeEndpoint_contentType :: Lens' InvokeEndpoint (Maybe Text)

The MIME type of the input data in the request body.

invokeEndpoint_targetContainerHostname :: Lens' InvokeEndpoint (Maybe Text)

If the endpoint hosts multiple containers and is configured to use direct invocation, this parameter specifies the host name of the container to invoke.

invokeEndpoint_endpointName :: Lens' InvokeEndpoint Text

The name of the endpoint that you specified when you created the endpoint using the CreateEndpoint API.

invokeEndpoint_body :: Lens' InvokeEndpoint Text

Provides input data, in the format specified in the ContentType request header. Amazon SageMaker passes all of the data in the body to the model.

For information about the format of the request body, see Common Data Formats-Inference.

Destructuring the Response

data InvokeEndpointResponse

See: newInvokeEndpointResponse smart constructor.

Constructors

InvokeEndpointResponse'

Fields

invokedProductionVariant :: Maybe Text

Identifies the production variant that was invoked.

customAttributes :: Maybe (Sensitive Text)

Provides additional information in the response about the inference returned by a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to return an ID received in the CustomAttributes header of a request or other metadata that a service endpoint was programmed to produce. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1). If the customer wants the custom attribute returned, the model must set the custom attribute to be included on the way back.

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

contentType :: Maybe Text

The MIME type of the inference returned in the response body.

httpStatus :: Int

The response's http status code.

body :: Text

Includes the inference provided by the model.

For information about the format of the response body, see Common Data Formats-Inference.

newInvokeEndpointResponse

Arguments

:: Int
$sel:httpStatus:InvokeEndpointResponse'

-> Text
$sel:body:InvokeEndpoint'

-> InvokeEndpointResponse

Create a value of InvokeEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invokedProductionVariant:InvokeEndpointResponse', invokeEndpointResponse_invokedProductionVariant - Identifies the production variant that was invoked.

$sel:customAttributes:InvokeEndpoint', invokeEndpointResponse_customAttributes - Provides additional information in the response about the inference returned by a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to return an ID received in the CustomAttributes header of a request or other metadata that a service endpoint was programmed to produce. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1). If the customer wants the custom attribute returned, the model must set the custom attribute to be included on the way back.

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

$sel:contentType:InvokeEndpoint', invokeEndpointResponse_contentType - The MIME type of the inference returned in the response body.

$sel:httpStatus:InvokeEndpointResponse', invokeEndpointResponse_httpStatus - The response's http status code.

$sel:body:InvokeEndpoint', invokeEndpointResponse_body - Includes the inference provided by the model.

For information about the format of the response body, see Common Data Formats-Inference.

Response Lenses

invokeEndpointResponse_invokedProductionVariant :: Lens' InvokeEndpointResponse (Maybe Text)

Identifies the production variant that was invoked.

invokeEndpointResponse_customAttributes :: Lens' InvokeEndpointResponse (Maybe Text)

Provides additional information in the response about the inference returned by a model hosted at an Amazon SageMaker endpoint. The information is an opaque value that is forwarded verbatim. You could use this value, for example, to return an ID received in the CustomAttributes header of a request or other metadata that a service endpoint was programmed to produce. The value must consist of no more than 1024 visible US-ASCII characters as specified in Section 3.3.6. Field Value Components of the Hypertext Transfer Protocol (HTTP/1.1). If the customer wants the custom attribute returned, the model must set the custom attribute to be included on the way back.

The code in your model is responsible for setting or updating any custom attributes in the response. If your code does not set this value in the response, an empty value is returned. For example, if a custom attribute represents the trace ID, your model can prepend the custom attribute with Trace ID: in your post-processing function.

This feature is currently supported in the AWS SDKs but not in the Amazon SageMaker Python SDK.

invokeEndpointResponse_contentType :: Lens' InvokeEndpointResponse (Maybe Text)

The MIME type of the inference returned in the response body.

invokeEndpointResponse_httpStatus :: Lens' InvokeEndpointResponse Int

The response's http status code.

invokeEndpointResponse_body :: Lens' InvokeEndpointResponse Text

Includes the inference provided by the model.

For information about the format of the response body, see Common Data Formats-Inference.