Amazonka.SavingsPlans.DescribeSavingsPlans

Description

Describes the specified Savings Plans.

Creating a Request

data DescribeSavingsPlans

See: newDescribeSavingsPlans smart constructor.

Constructors

DescribeSavingsPlans'

Fields

states :: Maybe [SavingsPlanState]

The states.

savingsPlanIds :: Maybe [Text]

The IDs of the Savings Plans.

filters :: Maybe [SavingsPlanFilter]

The filters.

nextToken :: Maybe Text

The token for the next page of results.

savingsPlanArns :: Maybe [Text]

The Amazon Resource Names (ARN) of the Savings Plans.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

newDescribeSavingsPlans :: DescribeSavingsPlans

Create a value of DescribeSavingsPlans with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:states:DescribeSavingsPlans', describeSavingsPlans_states - The states.

$sel:savingsPlanIds:DescribeSavingsPlans', describeSavingsPlans_savingsPlanIds - The IDs of the Savings Plans.

$sel:filters:DescribeSavingsPlans', describeSavingsPlans_filters - The filters.

$sel:nextToken:DescribeSavingsPlans', describeSavingsPlans_nextToken - The token for the next page of results.

$sel:savingsPlanArns:DescribeSavingsPlans', describeSavingsPlans_savingsPlanArns - The Amazon Resource Names (ARN) of the Savings Plans.

$sel:maxResults:DescribeSavingsPlans', describeSavingsPlans_maxResults - The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

Request Lenses

describeSavingsPlans_states :: Lens' DescribeSavingsPlans (Maybe [SavingsPlanState])

The states.

describeSavingsPlans_savingsPlanIds :: Lens' DescribeSavingsPlans (Maybe [Text])

The IDs of the Savings Plans.

describeSavingsPlans_filters :: Lens' DescribeSavingsPlans (Maybe [SavingsPlanFilter])

The filters.

describeSavingsPlans_nextToken :: Lens' DescribeSavingsPlans (Maybe Text)

The token for the next page of results.

describeSavingsPlans_savingsPlanArns :: Lens' DescribeSavingsPlans (Maybe [Text])

The Amazon Resource Names (ARN) of the Savings Plans.

describeSavingsPlans_maxResults :: Lens' DescribeSavingsPlans (Maybe Natural)

The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

Destructuring the Response

data DescribeSavingsPlansResponse

See: newDescribeSavingsPlansResponse smart constructor.

Constructors

DescribeSavingsPlansResponse'

Fields

savingsPlans :: Maybe [SavingsPlan]

Information about the Savings Plans.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeSavingsPlansResponse

Arguments

:: Int
$sel:httpStatus:DescribeSavingsPlansResponse'

-> DescribeSavingsPlansResponse

Create a value of DescribeSavingsPlansResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:savingsPlans:DescribeSavingsPlansResponse', describeSavingsPlansResponse_savingsPlans - Information about the Savings Plans.

$sel:nextToken:DescribeSavingsPlans', describeSavingsPlansResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeSavingsPlansResponse', describeSavingsPlansResponse_httpStatus - The response's http status code.

Response Lenses

describeSavingsPlansResponse_savingsPlans :: Lens' DescribeSavingsPlansResponse (Maybe [SavingsPlan])

Information about the Savings Plans.

describeSavingsPlansResponse_nextToken :: Lens' DescribeSavingsPlansResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeSavingsPlansResponse_httpStatus :: Lens' DescribeSavingsPlansResponse Int

The response's http status code.