Amazonka.SavingsPlans.DescribeSavingsPlansOfferings

Description

Describes the specified Savings Plans offerings.

Creating a Request

data DescribeSavingsPlansOfferings

See: newDescribeSavingsPlansOfferings smart constructor.

Constructors

DescribeSavingsPlansOfferings'

Fields

serviceCodes :: Maybe [Text]

The services.

productType :: Maybe SavingsPlanProductType

The product type.

filters :: Maybe [SavingsPlanOfferingFilterElement]

The filters.

offeringIds :: Maybe [Text]

The IDs of the offerings.

currencies :: Maybe [CurrencyCode]

The currencies.

nextToken :: Maybe Text

The token for the next page of results.

paymentOptions :: Maybe [SavingsPlanPaymentOption]

The payment options.

descriptions :: Maybe [Text]

The descriptions.

durations :: Maybe [Natural]

The durations, in seconds.

planTypes :: Maybe [SavingsPlanType]

The plan type.

usageTypes :: Maybe [Text]

The usage details of the line item in the billing report.

operations :: Maybe [Text]

The specific AWS operation for the line item in the billing report.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

newDescribeSavingsPlansOfferings :: DescribeSavingsPlansOfferings

Create a value of DescribeSavingsPlansOfferings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceCodes:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_serviceCodes - The services.

$sel:productType:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_productType - The product type.

$sel:filters:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_filters - The filters.

$sel:offeringIds:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_offeringIds - The IDs of the offerings.

$sel:currencies:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_currencies - The currencies.

$sel:nextToken:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_nextToken - The token for the next page of results.

$sel:paymentOptions:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_paymentOptions - The payment options.

$sel:descriptions:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_descriptions - The descriptions.

$sel:durations:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_durations - The durations, in seconds.

$sel:planTypes:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_planTypes - The plan type.

$sel:usageTypes:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_usageTypes - The usage details of the line item in the billing report.

$sel:operations:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_operations - The specific AWS operation for the line item in the billing report.

$sel:maxResults:DescribeSavingsPlansOfferings', describeSavingsPlansOfferings_maxResults - The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

Request Lenses

describeSavingsPlansOfferings_serviceCodes :: Lens' DescribeSavingsPlansOfferings (Maybe [Text])

The services.

describeSavingsPlansOfferings_productType :: Lens' DescribeSavingsPlansOfferings (Maybe SavingsPlanProductType)

The product type.

describeSavingsPlansOfferings_filters :: Lens' DescribeSavingsPlansOfferings (Maybe [SavingsPlanOfferingFilterElement])

The filters.

describeSavingsPlansOfferings_offeringIds :: Lens' DescribeSavingsPlansOfferings (Maybe [Text])

The IDs of the offerings.

describeSavingsPlansOfferings_currencies :: Lens' DescribeSavingsPlansOfferings (Maybe [CurrencyCode])

The currencies.

describeSavingsPlansOfferings_nextToken :: Lens' DescribeSavingsPlansOfferings (Maybe Text)

The token for the next page of results.

describeSavingsPlansOfferings_paymentOptions :: Lens' DescribeSavingsPlansOfferings (Maybe [SavingsPlanPaymentOption])

The payment options.

describeSavingsPlansOfferings_descriptions :: Lens' DescribeSavingsPlansOfferings (Maybe [Text])

The descriptions.

describeSavingsPlansOfferings_durations :: Lens' DescribeSavingsPlansOfferings (Maybe [Natural])

The durations, in seconds.

describeSavingsPlansOfferings_planTypes :: Lens' DescribeSavingsPlansOfferings (Maybe [SavingsPlanType])

The plan type.

describeSavingsPlansOfferings_usageTypes :: Lens' DescribeSavingsPlansOfferings (Maybe [Text])

The usage details of the line item in the billing report.

describeSavingsPlansOfferings_operations :: Lens' DescribeSavingsPlansOfferings (Maybe [Text])

The specific AWS operation for the line item in the billing report.

describeSavingsPlansOfferings_maxResults :: Lens' DescribeSavingsPlansOfferings (Maybe Natural)

The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

Destructuring the Response

data DescribeSavingsPlansOfferingsResponse

See: newDescribeSavingsPlansOfferingsResponse smart constructor.

Constructors

DescribeSavingsPlansOfferingsResponse'

Fields

searchResults :: Maybe [SavingsPlanOffering]

Information about the Savings Plans offerings.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newDescribeSavingsPlansOfferingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeSavingsPlansOfferingsResponse'

-> DescribeSavingsPlansOfferingsResponse

Create a value of DescribeSavingsPlansOfferingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:searchResults:DescribeSavingsPlansOfferingsResponse', describeSavingsPlansOfferingsResponse_searchResults - Information about the Savings Plans offerings.

$sel:nextToken:DescribeSavingsPlansOfferings', describeSavingsPlansOfferingsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:DescribeSavingsPlansOfferingsResponse', describeSavingsPlansOfferingsResponse_httpStatus - The response's http status code.

Response Lenses

describeSavingsPlansOfferingsResponse_searchResults :: Lens' DescribeSavingsPlansOfferingsResponse (Maybe [SavingsPlanOffering])

Information about the Savings Plans offerings.

describeSavingsPlansOfferingsResponse_nextToken :: Lens' DescribeSavingsPlansOfferingsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeSavingsPlansOfferingsResponse_httpStatus :: Lens' DescribeSavingsPlansOfferingsResponse Int

The response's http status code.