Amazonka.SavingsPlans.Lens

Description

Operations

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

Information about the tags.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DescribeSavingsPlanRates

describeSavingsPlanRates_filters :: Lens' DescribeSavingsPlanRates (Maybe [SavingsPlanRateFilter])

The filters.

describeSavingsPlanRates_nextToken :: Lens' DescribeSavingsPlanRates (Maybe Text)

The token for the next page of results.

describeSavingsPlanRates_maxResults :: Lens' DescribeSavingsPlanRates (Maybe Natural)

The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

describeSavingsPlanRates_savingsPlanId :: Lens' DescribeSavingsPlanRates Text

The ID of the Savings Plan.

describeSavingsPlanRatesResponse_searchResults :: Lens' DescribeSavingsPlanRatesResponse (Maybe [SavingsPlanRate])

Information about the Savings Plans rates.

describeSavingsPlanRatesResponse_savingsPlanId :: Lens' DescribeSavingsPlanRatesResponse (Maybe Text)

The ID of the Savings Plan.

describeSavingsPlanRatesResponse_nextToken :: Lens' DescribeSavingsPlanRatesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeSavingsPlanRatesResponse_httpStatus :: Lens' DescribeSavingsPlanRatesResponse Int

The response's http status code.

DeleteQueuedSavingsPlan

deleteQueuedSavingsPlan_savingsPlanId :: Lens' DeleteQueuedSavingsPlan Text

The ID of the Savings Plan.

deleteQueuedSavingsPlanResponse_httpStatus :: Lens' DeleteQueuedSavingsPlanResponse Int

The response's http status code.

CreateSavingsPlan

createSavingsPlan_clientToken :: Lens' CreateSavingsPlan (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createSavingsPlan_purchaseTime :: Lens' CreateSavingsPlan (Maybe UTCTime)

The time at which to purchase the Savings Plan, in UTC format (YYYY-MM-DDTHH:MM:SSZ).

createSavingsPlan_upfrontPaymentAmount :: Lens' CreateSavingsPlan (Maybe Text)

The up-front payment amount. This is a whole number between 50 and 99 percent of the total value of the Savings Plan. This parameter is supported only if the payment option is Partial Upfront.

createSavingsPlan_tags :: Lens' CreateSavingsPlan (Maybe (HashMap Text Text))

One or more tags.

createSavingsPlan_savingsPlanOfferingId :: Lens' CreateSavingsPlan Text

The ID of the offering.

createSavingsPlan_commitment :: Lens' CreateSavingsPlan Text

The hourly commitment, in USD. This is a value between 0.001 and 1 million. You cannot specify more than three digits after the decimal point.

createSavingsPlanResponse_savingsPlanId :: Lens' CreateSavingsPlanResponse (Maybe Text)

The ID of the Savings Plan.

createSavingsPlanResponse_httpStatus :: Lens' CreateSavingsPlanResponse Int

The response's http status code.

DescribeSavingsPlansOfferings

describeSavingsPlansOfferings_serviceCodes :: Lens' DescribeSavingsPlansOfferings (Maybe [Text])

The services.

describeSavingsPlansOfferings_productType :: Lens' DescribeSavingsPlansOfferings (Maybe SavingsPlanProductType)

The product type.

describeSavingsPlansOfferings_filters :: Lens' DescribeSavingsPlansOfferings (Maybe [SavingsPlanOfferingFilterElement])

The filters.

describeSavingsPlansOfferings_offeringIds :: Lens' DescribeSavingsPlansOfferings (Maybe [Text])

The IDs of the offerings.

describeSavingsPlansOfferings_currencies :: Lens' DescribeSavingsPlansOfferings (Maybe [CurrencyCode])

The currencies.

describeSavingsPlansOfferings_nextToken :: Lens' DescribeSavingsPlansOfferings (Maybe Text)

The token for the next page of results.

describeSavingsPlansOfferings_paymentOptions :: Lens' DescribeSavingsPlansOfferings (Maybe [SavingsPlanPaymentOption])

The payment options.

describeSavingsPlansOfferings_descriptions :: Lens' DescribeSavingsPlansOfferings (Maybe [Text])

The descriptions.

describeSavingsPlansOfferings_durations :: Lens' DescribeSavingsPlansOfferings (Maybe [Natural])

The durations, in seconds.

describeSavingsPlansOfferings_planTypes :: Lens' DescribeSavingsPlansOfferings (Maybe [SavingsPlanType])

The plan type.

describeSavingsPlansOfferings_usageTypes :: Lens' DescribeSavingsPlansOfferings (Maybe [Text])

The usage details of the line item in the billing report.

describeSavingsPlansOfferings_operations :: Lens' DescribeSavingsPlansOfferings (Maybe [Text])

The specific AWS operation for the line item in the billing report.

describeSavingsPlansOfferings_maxResults :: Lens' DescribeSavingsPlansOfferings (Maybe Natural)

The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

describeSavingsPlansOfferingsResponse_searchResults :: Lens' DescribeSavingsPlansOfferingsResponse (Maybe [SavingsPlanOffering])

Information about the Savings Plans offerings.

describeSavingsPlansOfferingsResponse_nextToken :: Lens' DescribeSavingsPlansOfferingsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeSavingsPlansOfferingsResponse_httpStatus :: Lens' DescribeSavingsPlansOfferingsResponse Int

The response's http status code.

DescribeSavingsPlans

describeSavingsPlans_states :: Lens' DescribeSavingsPlans (Maybe [SavingsPlanState])

The states.

describeSavingsPlans_savingsPlanIds :: Lens' DescribeSavingsPlans (Maybe [Text])

The IDs of the Savings Plans.

describeSavingsPlans_filters :: Lens' DescribeSavingsPlans (Maybe [SavingsPlanFilter])

The filters.

describeSavingsPlans_nextToken :: Lens' DescribeSavingsPlans (Maybe Text)

The token for the next page of results.

describeSavingsPlans_savingsPlanArns :: Lens' DescribeSavingsPlans (Maybe [Text])

The Amazon Resource Names (ARN) of the Savings Plans.

describeSavingsPlans_maxResults :: Lens' DescribeSavingsPlans (Maybe Natural)

The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

describeSavingsPlansResponse_savingsPlans :: Lens' DescribeSavingsPlansResponse (Maybe [SavingsPlan])

Information about the Savings Plans.

describeSavingsPlansResponse_nextToken :: Lens' DescribeSavingsPlansResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeSavingsPlansResponse_httpStatus :: Lens' DescribeSavingsPlansResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

One or more tags. For example, { "tags": {"key1":"value1", "key2":"value2"} }.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource.

untagResource_tagKeys :: Lens' UntagResource [Text]

The tag keys.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DescribeSavingsPlansOfferingRates

describeSavingsPlansOfferingRates_savingsPlanOfferingIds :: Lens' DescribeSavingsPlansOfferingRates (Maybe [Text])

The IDs of the offerings.

describeSavingsPlansOfferingRates_serviceCodes :: Lens' DescribeSavingsPlansOfferingRates (Maybe [SavingsPlanRateServiceCode])

The services.

describeSavingsPlansOfferingRates_filters :: Lens' DescribeSavingsPlansOfferingRates (Maybe [SavingsPlanOfferingRateFilterElement])

The filters.

describeSavingsPlansOfferingRates_nextToken :: Lens' DescribeSavingsPlansOfferingRates (Maybe Text)

The token for the next page of results.

describeSavingsPlansOfferingRates_savingsPlanTypes :: Lens' DescribeSavingsPlansOfferingRates (Maybe [SavingsPlanType])

The plan types.

describeSavingsPlansOfferingRates_products :: Lens' DescribeSavingsPlansOfferingRates (Maybe [SavingsPlanProductType])

The AWS products.

describeSavingsPlansOfferingRates_savingsPlanPaymentOptions :: Lens' DescribeSavingsPlansOfferingRates (Maybe [SavingsPlanPaymentOption])

The payment options.

describeSavingsPlansOfferingRates_usageTypes :: Lens' DescribeSavingsPlansOfferingRates (Maybe [Text])

The usage details of the line item in the billing report.

describeSavingsPlansOfferingRates_operations :: Lens' DescribeSavingsPlansOfferingRates (Maybe [Text])

The specific AWS operation for the line item in the billing report.

describeSavingsPlansOfferingRates_maxResults :: Lens' DescribeSavingsPlansOfferingRates (Maybe Natural)

The maximum number of results to return with a single call. To retrieve additional results, make another call with the returned token value.

describeSavingsPlansOfferingRatesResponse_searchResults :: Lens' DescribeSavingsPlansOfferingRatesResponse (Maybe [SavingsPlanOfferingRate])

Information about the Savings Plans offering rates.

describeSavingsPlansOfferingRatesResponse_nextToken :: Lens' DescribeSavingsPlansOfferingRatesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

describeSavingsPlansOfferingRatesResponse_httpStatus :: Lens' DescribeSavingsPlansOfferingRatesResponse Int

The response's http status code.

Types

ParentSavingsPlanOffering

parentSavingsPlanOffering_currency :: Lens' ParentSavingsPlanOffering (Maybe CurrencyCode)

The currency.

parentSavingsPlanOffering_durationSeconds :: Lens' ParentSavingsPlanOffering (Maybe Natural)

The duration, in seconds.

parentSavingsPlanOffering_offeringId :: Lens' ParentSavingsPlanOffering (Maybe Text)

The ID of the offering.

parentSavingsPlanOffering_planDescription :: Lens' ParentSavingsPlanOffering (Maybe Text)

The description.

parentSavingsPlanOffering_planType :: Lens' ParentSavingsPlanOffering (Maybe SavingsPlanType)

The plan type.

parentSavingsPlanOffering_paymentOption :: Lens' ParentSavingsPlanOffering (Maybe SavingsPlanPaymentOption)

The payment option.

SavingsPlan

savingsPlan_termDurationInSeconds :: Lens' SavingsPlan (Maybe Integer)

The duration of the term, in seconds.

savingsPlan_savingsPlanType :: Lens' SavingsPlan (Maybe SavingsPlanType)

The plan type.

savingsPlan_state :: Lens' SavingsPlan (Maybe SavingsPlanState)

The state.

savingsPlan_productTypes :: Lens' SavingsPlan (Maybe [SavingsPlanProductType])

The product types.

savingsPlan_start :: Lens' SavingsPlan (Maybe Text)

The start time.

savingsPlan_ec2InstanceFamily :: Lens' SavingsPlan (Maybe Text)

The EC2 instance family.

savingsPlan_savingsPlanId :: Lens' SavingsPlan (Maybe Text)

The ID of the Savings Plan.

savingsPlan_recurringPaymentAmount :: Lens' SavingsPlan (Maybe Text)

The recurring payment amount.

savingsPlan_currency :: Lens' SavingsPlan (Maybe CurrencyCode)

The currency.

savingsPlan_end :: Lens' SavingsPlan (Maybe Text)

The end time.

savingsPlan_upfrontPaymentAmount :: Lens' SavingsPlan (Maybe Text)

The up-front payment amount.

savingsPlan_offeringId :: Lens' SavingsPlan (Maybe Text)

The ID of the offering.

savingsPlan_region :: Lens' SavingsPlan (Maybe Text)

The AWS Region.

savingsPlan_commitment :: Lens' SavingsPlan (Maybe Text)

The hourly commitment, in USD.

savingsPlan_description :: Lens' SavingsPlan (Maybe Text)

The description.

savingsPlan_paymentOption :: Lens' SavingsPlan (Maybe SavingsPlanPaymentOption)

The payment option.

savingsPlan_savingsPlanArn :: Lens' SavingsPlan (Maybe Text)

The Amazon Resource Name (ARN) of the Savings Plan.

savingsPlan_tags :: Lens' SavingsPlan (Maybe (HashMap Text Text))

One or more tags.

SavingsPlanFilter

savingsPlanFilter_values :: Lens' SavingsPlanFilter (Maybe [Text])

The filter value.

savingsPlanFilter_name :: Lens' SavingsPlanFilter (Maybe SavingsPlansFilterName)

The filter name.

SavingsPlanOffering

savingsPlanOffering_operation :: Lens' SavingsPlanOffering (Maybe Text)

The specific AWS operation for the line item in the billing report.

savingsPlanOffering_usageType :: Lens' SavingsPlanOffering (Maybe Text)

The usage details of the line item in the billing report.

savingsPlanOffering_productTypes :: Lens' SavingsPlanOffering (Maybe [SavingsPlanProductType])

The product type.

savingsPlanOffering_currency :: Lens' SavingsPlanOffering (Maybe CurrencyCode)

The currency.

savingsPlanOffering_durationSeconds :: Lens' SavingsPlanOffering (Maybe Natural)

The duration, in seconds.

savingsPlanOffering_offeringId :: Lens' SavingsPlanOffering (Maybe Text)

The ID of the offering.

savingsPlanOffering_serviceCode :: Lens' SavingsPlanOffering (Maybe Text)

The service.

savingsPlanOffering_planType :: Lens' SavingsPlanOffering (Maybe SavingsPlanType)

The plan type.

savingsPlanOffering_description :: Lens' SavingsPlanOffering (Maybe Text)

The description.

savingsPlanOffering_paymentOption :: Lens' SavingsPlanOffering (Maybe SavingsPlanPaymentOption)

The payment option.

savingsPlanOffering_properties :: Lens' SavingsPlanOffering (Maybe [SavingsPlanOfferingProperty])

The properties.

SavingsPlanOfferingFilterElement

savingsPlanOfferingFilterElement_values :: Lens' SavingsPlanOfferingFilterElement (Maybe [Text])

The filter values.

savingsPlanOfferingFilterElement_name :: Lens' SavingsPlanOfferingFilterElement (Maybe SavingsPlanOfferingFilterAttribute)

The filter name.

SavingsPlanOfferingProperty

savingsPlanOfferingProperty_value :: Lens' SavingsPlanOfferingProperty (Maybe Text)

The property value.

savingsPlanOfferingProperty_name :: Lens' SavingsPlanOfferingProperty (Maybe SavingsPlanOfferingPropertyKey)

The property name.

SavingsPlanOfferingRate

savingsPlanOfferingRate_operation :: Lens' SavingsPlanOfferingRate (Maybe Text)

The specific AWS operation for the line item in the billing report.

savingsPlanOfferingRate_usageType :: Lens' SavingsPlanOfferingRate (Maybe Text)

The usage details of the line item in the billing report.

savingsPlanOfferingRate_productType :: Lens' SavingsPlanOfferingRate (Maybe SavingsPlanProductType)

The product type.

savingsPlanOfferingRate_savingsPlanOffering :: Lens' SavingsPlanOfferingRate (Maybe ParentSavingsPlanOffering)

The Savings Plan offering.

savingsPlanOfferingRate_rate :: Lens' SavingsPlanOfferingRate (Maybe Text)

The Savings Plan rate.

savingsPlanOfferingRate_serviceCode :: Lens' SavingsPlanOfferingRate (Maybe SavingsPlanRateServiceCode)

The service.

savingsPlanOfferingRate_unit :: Lens' SavingsPlanOfferingRate (Maybe SavingsPlanRateUnit)

The unit.

savingsPlanOfferingRate_properties :: Lens' SavingsPlanOfferingRate (Maybe [SavingsPlanOfferingRateProperty])

The properties.

SavingsPlanOfferingRateFilterElement

savingsPlanOfferingRateFilterElement_values :: Lens' SavingsPlanOfferingRateFilterElement (Maybe [Text])

The filter values.

savingsPlanOfferingRateFilterElement_name :: Lens' SavingsPlanOfferingRateFilterElement (Maybe SavingsPlanRateFilterAttribute)

The filter name.

SavingsPlanOfferingRateProperty

savingsPlanOfferingRateProperty_value :: Lens' SavingsPlanOfferingRateProperty (Maybe Text)

The property value.

savingsPlanOfferingRateProperty_name :: Lens' SavingsPlanOfferingRateProperty (Maybe Text)

The property name.

SavingsPlanRate

savingsPlanRate_operation :: Lens' SavingsPlanRate (Maybe Text)

The specific AWS operation for the line item in the billing report.

savingsPlanRate_usageType :: Lens' SavingsPlanRate (Maybe Text)

The usage details of the line item in the billing report.

savingsPlanRate_productType :: Lens' SavingsPlanRate (Maybe SavingsPlanProductType)

The product type.

savingsPlanRate_currency :: Lens' SavingsPlanRate (Maybe CurrencyCode)

The currency.

savingsPlanRate_rate :: Lens' SavingsPlanRate (Maybe Text)

The rate.

savingsPlanRate_serviceCode :: Lens' SavingsPlanRate (Maybe SavingsPlanRateServiceCode)

The service.

savingsPlanRate_unit :: Lens' SavingsPlanRate (Maybe SavingsPlanRateUnit)

The unit.

savingsPlanRate_properties :: Lens' SavingsPlanRate (Maybe [SavingsPlanRateProperty])

The properties.

SavingsPlanRateFilter

savingsPlanRateFilter_values :: Lens' SavingsPlanRateFilter (Maybe [Text])

The filter values.

savingsPlanRateFilter_name :: Lens' SavingsPlanRateFilter (Maybe SavingsPlanRateFilterName)

The filter name.

SavingsPlanRateProperty

savingsPlanRateProperty_value :: Lens' SavingsPlanRateProperty (Maybe Text)

The property value.

savingsPlanRateProperty_name :: Lens' SavingsPlanRateProperty (Maybe SavingsPlanRatePropertyKey)

The property name.