Amazonka.SavingsPlans.Types.SavingsPlanOffering

Description

Documentation

data SavingsPlanOffering

Information about a Savings Plan offering.

See: newSavingsPlanOffering smart constructor.

Constructors

SavingsPlanOffering'

Fields

operation :: Maybe Text

The specific AWS operation for the line item in the billing report.

usageType :: Maybe Text

The usage details of the line item in the billing report.

productTypes :: Maybe [SavingsPlanProductType]

The product type.

currency :: Maybe CurrencyCode

The currency.

durationSeconds :: Maybe Natural

The duration, in seconds.

offeringId :: Maybe Text

The ID of the offering.

serviceCode :: Maybe Text

The service.

planType :: Maybe SavingsPlanType

The plan type.

description :: Maybe Text

The description.

paymentOption :: Maybe SavingsPlanPaymentOption

The payment option.

properties :: Maybe [SavingsPlanOfferingProperty]

The properties.

newSavingsPlanOffering :: SavingsPlanOffering

Create a value of SavingsPlanOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:SavingsPlanOffering', savingsPlanOffering_operation - The specific AWS operation for the line item in the billing report.

$sel:usageType:SavingsPlanOffering', savingsPlanOffering_usageType - The usage details of the line item in the billing report.

$sel:productTypes:SavingsPlanOffering', savingsPlanOffering_productTypes - The product type.

$sel:currency:SavingsPlanOffering', savingsPlanOffering_currency - The currency.

$sel:durationSeconds:SavingsPlanOffering', savingsPlanOffering_durationSeconds - The duration, in seconds.

$sel:offeringId:SavingsPlanOffering', savingsPlanOffering_offeringId - The ID of the offering.

$sel:serviceCode:SavingsPlanOffering', savingsPlanOffering_serviceCode - The service.

$sel:planType:SavingsPlanOffering', savingsPlanOffering_planType - The plan type.

$sel:description:SavingsPlanOffering', savingsPlanOffering_description - The description.

$sel:paymentOption:SavingsPlanOffering', savingsPlanOffering_paymentOption - The payment option.

$sel:properties:SavingsPlanOffering', savingsPlanOffering_properties - The properties.

savingsPlanOffering_operation :: Lens' SavingsPlanOffering (Maybe Text)

The specific AWS operation for the line item in the billing report.

savingsPlanOffering_usageType :: Lens' SavingsPlanOffering (Maybe Text)

The usage details of the line item in the billing report.

savingsPlanOffering_productTypes :: Lens' SavingsPlanOffering (Maybe [SavingsPlanProductType])

The product type.

savingsPlanOffering_currency :: Lens' SavingsPlanOffering (Maybe CurrencyCode)

The currency.

savingsPlanOffering_durationSeconds :: Lens' SavingsPlanOffering (Maybe Natural)

The duration, in seconds.

savingsPlanOffering_offeringId :: Lens' SavingsPlanOffering (Maybe Text)

The ID of the offering.

savingsPlanOffering_serviceCode :: Lens' SavingsPlanOffering (Maybe Text)

The service.

savingsPlanOffering_planType :: Lens' SavingsPlanOffering (Maybe SavingsPlanType)

The plan type.

savingsPlanOffering_description :: Lens' SavingsPlanOffering (Maybe Text)

The description.

savingsPlanOffering_paymentOption :: Lens' SavingsPlanOffering (Maybe SavingsPlanPaymentOption)

The payment option.

savingsPlanOffering_properties :: Lens' SavingsPlanOffering (Maybe [SavingsPlanOfferingProperty])

The properties.