Amazonka.SavingsPlans.Types.SavingsPlanRate

Description

Documentation

data SavingsPlanRate

Information about a Savings Plan rate.

See: newSavingsPlanRate smart constructor.

Constructors

SavingsPlanRate'

Fields

operation :: Maybe Text

The specific AWS operation for the line item in the billing report.

usageType :: Maybe Text

The usage details of the line item in the billing report.

productType :: Maybe SavingsPlanProductType

The product type.

currency :: Maybe CurrencyCode

The currency.

rate :: Maybe Text

The rate.

serviceCode :: Maybe SavingsPlanRateServiceCode

The service.

unit :: Maybe SavingsPlanRateUnit

The unit.

properties :: Maybe [SavingsPlanRateProperty]

The properties.

newSavingsPlanRate :: SavingsPlanRate

Create a value of SavingsPlanRate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:SavingsPlanRate', savingsPlanRate_operation - The specific AWS operation for the line item in the billing report.

$sel:usageType:SavingsPlanRate', savingsPlanRate_usageType - The usage details of the line item in the billing report.

$sel:productType:SavingsPlanRate', savingsPlanRate_productType - The product type.

$sel:currency:SavingsPlanRate', savingsPlanRate_currency - The currency.

$sel:rate:SavingsPlanRate', savingsPlanRate_rate - The rate.

$sel:serviceCode:SavingsPlanRate', savingsPlanRate_serviceCode - The service.

$sel:unit:SavingsPlanRate', savingsPlanRate_unit - The unit.

$sel:properties:SavingsPlanRate', savingsPlanRate_properties - The properties.

savingsPlanRate_operation :: Lens' SavingsPlanRate (Maybe Text)

The specific AWS operation for the line item in the billing report.

savingsPlanRate_usageType :: Lens' SavingsPlanRate (Maybe Text)

The usage details of the line item in the billing report.

savingsPlanRate_productType :: Lens' SavingsPlanRate (Maybe SavingsPlanProductType)

The product type.

savingsPlanRate_currency :: Lens' SavingsPlanRate (Maybe CurrencyCode)

The currency.

savingsPlanRate_rate :: Lens' SavingsPlanRate (Maybe Text)

The rate.

savingsPlanRate_serviceCode :: Lens' SavingsPlanRate (Maybe SavingsPlanRateServiceCode)

The service.

savingsPlanRate_unit :: Lens' SavingsPlanRate (Maybe SavingsPlanRateUnit)

The unit.

savingsPlanRate_properties :: Lens' SavingsPlanRate (Maybe [SavingsPlanRateProperty])

The properties.