Amazonka.Schemas.DescribeCodeBinding

Description

Describe the code binding URI.

Creating a Request

data DescribeCodeBinding

See: newDescribeCodeBinding smart constructor.

Constructors

DescribeCodeBinding'

Fields

schemaVersion :: Maybe Text

Specifying this limits the results to only this schema version.

registryName :: Text

The name of the registry.

schemaName :: Text

The name of the schema.

language :: Text

The language of the code binding.

newDescribeCodeBinding

Arguments

:: Text
$sel:registryName:DescribeCodeBinding'

-> Text
$sel:schemaName:DescribeCodeBinding'

-> Text
$sel:language:DescribeCodeBinding'

-> DescribeCodeBinding

Create a value of DescribeCodeBinding with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersion:DescribeCodeBinding', describeCodeBinding_schemaVersion - Specifying this limits the results to only this schema version.

$sel:registryName:DescribeCodeBinding', describeCodeBinding_registryName - The name of the registry.

$sel:schemaName:DescribeCodeBinding', describeCodeBinding_schemaName - The name of the schema.

$sel:language:DescribeCodeBinding', describeCodeBinding_language - The language of the code binding.

Request Lenses

describeCodeBinding_schemaVersion :: Lens' DescribeCodeBinding (Maybe Text)

Specifying this limits the results to only this schema version.

describeCodeBinding_registryName :: Lens' DescribeCodeBinding Text

The name of the registry.

describeCodeBinding_schemaName :: Lens' DescribeCodeBinding Text

The name of the schema.

describeCodeBinding_language :: Lens' DescribeCodeBinding Text

The language of the code binding.

Destructuring the Response

data DescribeCodeBindingResponse

See: newDescribeCodeBindingResponse smart constructor.

Constructors

DescribeCodeBindingResponse'

Fields

status :: Maybe CodeGenerationStatus

The current status of code binding generation.

schemaVersion :: Maybe Text

The version number of the schema.

creationDate :: Maybe POSIX

The time and date that the code binding was created.

lastModified :: Maybe POSIX

The date and time that code bindings were modified.

httpStatus :: Int

The response's http status code.

newDescribeCodeBindingResponse

Arguments

:: Int
$sel:httpStatus:DescribeCodeBindingResponse'

-> DescribeCodeBindingResponse

Create a value of DescribeCodeBindingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeCodeBindingResponse', describeCodeBindingResponse_status - The current status of code binding generation.

$sel:schemaVersion:DescribeCodeBinding', describeCodeBindingResponse_schemaVersion - The version number of the schema.

$sel:creationDate:DescribeCodeBindingResponse', describeCodeBindingResponse_creationDate - The time and date that the code binding was created.

$sel:lastModified:DescribeCodeBindingResponse', describeCodeBindingResponse_lastModified - The date and time that code bindings were modified.

$sel:httpStatus:DescribeCodeBindingResponse', describeCodeBindingResponse_httpStatus - The response's http status code.

Response Lenses

describeCodeBindingResponse_status :: Lens' DescribeCodeBindingResponse (Maybe CodeGenerationStatus)

The current status of code binding generation.

describeCodeBindingResponse_schemaVersion :: Lens' DescribeCodeBindingResponse (Maybe Text)

The version number of the schema.

describeCodeBindingResponse_creationDate :: Lens' DescribeCodeBindingResponse (Maybe UTCTime)

The time and date that the code binding was created.

describeCodeBindingResponse_lastModified :: Lens' DescribeCodeBindingResponse (Maybe UTCTime)

The date and time that code bindings were modified.

describeCodeBindingResponse_httpStatus :: Lens' DescribeCodeBindingResponse Int

The response's http status code.