Amazonka.Schemas.DescribeSchema

Description

Retrieve the schema definition.

Creating a Request

data DescribeSchema

See: newDescribeSchema smart constructor.

Constructors

DescribeSchema'

Fields

schemaVersion :: Maybe Text

Specifying this limits the results to only this schema version.

registryName :: Text

The name of the registry.

schemaName :: Text

The name of the schema.

newDescribeSchema

Arguments

:: Text
$sel:registryName:DescribeSchema'

-> Text
$sel:schemaName:DescribeSchema'

-> DescribeSchema

Create a value of DescribeSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersion:DescribeSchema', describeSchema_schemaVersion - Specifying this limits the results to only this schema version.

$sel:registryName:DescribeSchema', describeSchema_registryName - The name of the registry.

$sel:schemaName:DescribeSchema', describeSchema_schemaName - The name of the schema.

Request Lenses

describeSchema_schemaVersion :: Lens' DescribeSchema (Maybe Text)

Specifying this limits the results to only this schema version.

describeSchema_registryName :: Lens' DescribeSchema Text

The name of the registry.

describeSchema_schemaName :: Lens' DescribeSchema Text

The name of the schema.

Destructuring the Response

data DescribeSchemaResponse

See: newDescribeSchemaResponse smart constructor.

Constructors

DescribeSchemaResponse'

Fields

schemaVersion :: Maybe Text

The version number of the schema

schemaName :: Maybe Text

The name of the schema.

content :: Maybe Text

The source of the schema definition.

schemaArn :: Maybe Text

The ARN of the schema.

type' :: Maybe Text

The type of the schema.

lastModified :: Maybe POSIX

The date and time that schema was modified.

description :: Maybe Text

The description of the schema.

versionCreatedDate :: Maybe POSIX

The date the schema version was created.

tags :: Maybe (HashMap Text Text)

Tags associated with the resource.

httpStatus :: Int

The response's http status code.

newDescribeSchemaResponse

Arguments

:: Int
$sel:httpStatus:DescribeSchemaResponse'

-> DescribeSchemaResponse

Create a value of DescribeSchemaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersion:DescribeSchema', describeSchemaResponse_schemaVersion - The version number of the schema

$sel:schemaName:DescribeSchema', describeSchemaResponse_schemaName - The name of the schema.

$sel:content:DescribeSchemaResponse', describeSchemaResponse_content - The source of the schema definition.

$sel:schemaArn:DescribeSchemaResponse', describeSchemaResponse_schemaArn - The ARN of the schema.

$sel:type':DescribeSchemaResponse', describeSchemaResponse_type - The type of the schema.

$sel:lastModified:DescribeSchemaResponse', describeSchemaResponse_lastModified - The date and time that schema was modified.

$sel:description:DescribeSchemaResponse', describeSchemaResponse_description - The description of the schema.

$sel:versionCreatedDate:DescribeSchemaResponse', describeSchemaResponse_versionCreatedDate - The date the schema version was created.

$sel:tags:DescribeSchemaResponse', describeSchemaResponse_tags - Tags associated with the resource.

$sel:httpStatus:DescribeSchemaResponse', describeSchemaResponse_httpStatus - The response's http status code.

Response Lenses

describeSchemaResponse_schemaVersion :: Lens' DescribeSchemaResponse (Maybe Text)

The version number of the schema

describeSchemaResponse_schemaName :: Lens' DescribeSchemaResponse (Maybe Text)

The name of the schema.

describeSchemaResponse_content :: Lens' DescribeSchemaResponse (Maybe Text)

The source of the schema definition.

describeSchemaResponse_schemaArn :: Lens' DescribeSchemaResponse (Maybe Text)

The ARN of the schema.

describeSchemaResponse_type :: Lens' DescribeSchemaResponse (Maybe Text)

The type of the schema.

describeSchemaResponse_lastModified :: Lens' DescribeSchemaResponse (Maybe UTCTime)

The date and time that schema was modified.

describeSchemaResponse_description :: Lens' DescribeSchemaResponse (Maybe Text)

The description of the schema.

describeSchemaResponse_versionCreatedDate :: Lens' DescribeSchemaResponse (Maybe UTCTime)

The date the schema version was created.

describeSchemaResponse_tags :: Lens' DescribeSchemaResponse (Maybe (HashMap Text Text))

Tags associated with the resource.

describeSchemaResponse_httpStatus :: Lens' DescribeSchemaResponse Int

The response's http status code.