Amazonka.Schemas.ExportSchema

Description

- | Undocumented operation.

Creating a Request

data ExportSchema

See: newExportSchema smart constructor.

Constructors

ExportSchema'

Fields

schemaVersion :: Maybe Text

Specifying this limits the results to only this schema version.

registryName :: Text

The name of the registry.

schemaName :: Text

The name of the schema.

type' :: Text

newExportSchema

Arguments

:: Text
$sel:registryName:ExportSchema'

-> Text
$sel:schemaName:ExportSchema'

-> Text
$sel:type':ExportSchema'

-> ExportSchema

Create a value of ExportSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersion:ExportSchema', exportSchema_schemaVersion - Specifying this limits the results to only this schema version.

$sel:registryName:ExportSchema', exportSchema_registryName - The name of the registry.

$sel:schemaName:ExportSchema', exportSchema_schemaName - The name of the schema.

$sel:type':ExportSchema', exportSchema_type - Undocumented member.

Request Lenses

exportSchema_schemaVersion :: Lens' ExportSchema (Maybe Text)

Specifying this limits the results to only this schema version.

exportSchema_registryName :: Lens' ExportSchema Text

The name of the registry.

exportSchema_schemaName :: Lens' ExportSchema Text

The name of the schema.

exportSchema_type :: Lens' ExportSchema Text

Undocumented member.

Destructuring the Response

data ExportSchemaResponse

See: newExportSchemaResponse smart constructor.

Constructors

ExportSchemaResponse'

Fields

schemaVersion :: Maybe Text

schemaName :: Maybe Text

content :: Maybe Text

schemaArn :: Maybe Text

type' :: Maybe Text

httpStatus :: Int

The response's http status code.

newExportSchemaResponse

Arguments

:: Int
$sel:httpStatus:ExportSchemaResponse'

-> ExportSchemaResponse

Create a value of ExportSchemaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersion:ExportSchema', exportSchemaResponse_schemaVersion - Undocumented member.

$sel:schemaName:ExportSchema', exportSchemaResponse_schemaName - Undocumented member.

$sel:content:ExportSchemaResponse', exportSchemaResponse_content - Undocumented member.

$sel:schemaArn:ExportSchemaResponse', exportSchemaResponse_schemaArn - Undocumented member.

$sel:type':ExportSchema', exportSchemaResponse_type - Undocumented member.

$sel:httpStatus:ExportSchemaResponse', exportSchemaResponse_httpStatus - The response's http status code.

Response Lenses

exportSchemaResponse_schemaVersion :: Lens' ExportSchemaResponse (Maybe Text)

Undocumented member.

exportSchemaResponse_schemaName :: Lens' ExportSchemaResponse (Maybe Text)

Undocumented member.

exportSchemaResponse_content :: Lens' ExportSchemaResponse (Maybe Text)

Undocumented member.

exportSchemaResponse_schemaArn :: Lens' ExportSchemaResponse (Maybe Text)

Undocumented member.

exportSchemaResponse_type :: Lens' ExportSchemaResponse (Maybe Text)

Undocumented member.

exportSchemaResponse_httpStatus :: Lens' ExportSchemaResponse Int

The response's http status code.