Amazonka.Schemas.GetCodeBindingSource

Description

Get the code binding source URI.

Creating a Request

data GetCodeBindingSource

See: newGetCodeBindingSource smart constructor.

Constructors

GetCodeBindingSource'

Fields

schemaVersion :: Maybe Text

Specifying this limits the results to only this schema version.

registryName :: Text

The name of the registry.

schemaName :: Text

The name of the schema.

language :: Text

The language of the code binding.

newGetCodeBindingSource

Arguments

:: Text
$sel:registryName:GetCodeBindingSource'

-> Text
$sel:schemaName:GetCodeBindingSource'

-> Text
$sel:language:GetCodeBindingSource'

-> GetCodeBindingSource

Create a value of GetCodeBindingSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersion:GetCodeBindingSource', getCodeBindingSource_schemaVersion - Specifying this limits the results to only this schema version.

$sel:registryName:GetCodeBindingSource', getCodeBindingSource_registryName - The name of the registry.

$sel:schemaName:GetCodeBindingSource', getCodeBindingSource_schemaName - The name of the schema.

$sel:language:GetCodeBindingSource', getCodeBindingSource_language - The language of the code binding.

Request Lenses

getCodeBindingSource_schemaVersion :: Lens' GetCodeBindingSource (Maybe Text)

Specifying this limits the results to only this schema version.

getCodeBindingSource_registryName :: Lens' GetCodeBindingSource Text

The name of the registry.

getCodeBindingSource_schemaName :: Lens' GetCodeBindingSource Text

The name of the schema.

getCodeBindingSource_language :: Lens' GetCodeBindingSource Text

The language of the code binding.

Destructuring the Response

data GetCodeBindingSourceResponse

See: newGetCodeBindingSourceResponse smart constructor.

Constructors

GetCodeBindingSourceResponse'

Fields

body :: Maybe ByteString

httpStatus :: Int

The response's http status code.

newGetCodeBindingSourceResponse

Arguments

:: Int
$sel:httpStatus:GetCodeBindingSourceResponse'

-> GetCodeBindingSourceResponse

Create a value of GetCodeBindingSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:body:GetCodeBindingSourceResponse', getCodeBindingSourceResponse_body - Undocumented member.

$sel:httpStatus:GetCodeBindingSourceResponse', getCodeBindingSourceResponse_httpStatus - The response's http status code.

Response Lenses

getCodeBindingSourceResponse_body :: Lens' GetCodeBindingSourceResponse (Maybe ByteString)

Undocumented member.

getCodeBindingSourceResponse_httpStatus :: Lens' GetCodeBindingSourceResponse Int

The response's http status code.