Amazonka.Schemas.SearchSchemas

Description

Search the schemas

This operation returns paginated results.

Creating a Request

data SearchSchemas

See: newSearchSchemas smart constructor.

Constructors

SearchSchemas'

Fields

nextToken :: Maybe Text

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

limit :: Maybe Int

registryName :: Text

The name of the registry.

keywords :: Text

Specifying this limits the results to only schemas that include the provided keywords.

newSearchSchemas

Arguments

:: Text
$sel:registryName:SearchSchemas'

-> Text
$sel:keywords:SearchSchemas'

-> SearchSchemas

Create a value of SearchSchemas with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchSchemas', searchSchemas_nextToken - The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

$sel:limit:SearchSchemas', searchSchemas_limit - Undocumented member.

$sel:registryName:SearchSchemas', searchSchemas_registryName - The name of the registry.

$sel:keywords:SearchSchemas', searchSchemas_keywords - Specifying this limits the results to only schemas that include the provided keywords.

Request Lenses

searchSchemas_nextToken :: Lens' SearchSchemas (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

searchSchemas_limit :: Lens' SearchSchemas (Maybe Int)

Undocumented member.

searchSchemas_registryName :: Lens' SearchSchemas Text

The name of the registry.

searchSchemas_keywords :: Lens' SearchSchemas Text

Specifying this limits the results to only schemas that include the provided keywords.

Destructuring the Response

data SearchSchemasResponse

See: newSearchSchemasResponse smart constructor.

Constructors

SearchSchemasResponse'

Fields

schemas :: Maybe [SearchSchemaSummary]

An array of SearchSchemaSummary information.

nextToken :: Maybe Text

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

httpStatus :: Int

The response's http status code.

newSearchSchemasResponse

Arguments

:: Int
$sel:httpStatus:SearchSchemasResponse'

-> SearchSchemasResponse

Create a value of SearchSchemasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemas:SearchSchemasResponse', searchSchemasResponse_schemas - An array of SearchSchemaSummary information.

$sel:nextToken:SearchSchemas', searchSchemasResponse_nextToken - The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

$sel:httpStatus:SearchSchemasResponse', searchSchemasResponse_httpStatus - The response's http status code.

Response Lenses

searchSchemasResponse_schemas :: Lens' SearchSchemasResponse (Maybe [SearchSchemaSummary])

An array of SearchSchemaSummary information.

searchSchemasResponse_nextToken :: Lens' SearchSchemasResponse (Maybe Text)

The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

searchSchemasResponse_httpStatus :: Lens' SearchSchemasResponse Int

The response's http status code.