Amazonka.Schemas.Types.DiscovererSummary

Description

Documentation

data DiscovererSummary

See: newDiscovererSummary smart constructor.

Constructors

DiscovererSummary'

Fields

state :: Maybe DiscovererState

The state of the discoverer.

crossAccount :: Maybe Bool

The Status if the discoverer will discover schemas from events sent from another account.

sourceArn :: Maybe Text

The ARN of the event bus.

discovererId :: Maybe Text

The ID of the discoverer.

tags :: Maybe (HashMap Text Text)

Tags associated with the resource.

discovererArn :: Maybe Text

The ARN of the discoverer.

newDiscovererSummary :: DiscovererSummary

Create a value of DiscovererSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:DiscovererSummary', discovererSummary_state - The state of the discoverer.

$sel:crossAccount:DiscovererSummary', discovererSummary_crossAccount - The Status if the discoverer will discover schemas from events sent from another account.

$sel:sourceArn:DiscovererSummary', discovererSummary_sourceArn - The ARN of the event bus.

$sel:discovererId:DiscovererSummary', discovererSummary_discovererId - The ID of the discoverer.

$sel:tags:DiscovererSummary', discovererSummary_tags - Tags associated with the resource.

$sel:discovererArn:DiscovererSummary', discovererSummary_discovererArn - The ARN of the discoverer.

discovererSummary_state :: Lens' DiscovererSummary (Maybe DiscovererState)

The state of the discoverer.

discovererSummary_crossAccount :: Lens' DiscovererSummary (Maybe Bool)

The Status if the discoverer will discover schemas from events sent from another account.

discovererSummary_sourceArn :: Lens' DiscovererSummary (Maybe Text)

The ARN of the event bus.

discovererSummary_discovererId :: Lens' DiscovererSummary (Maybe Text)

The ID of the discoverer.

discovererSummary_tags :: Lens' DiscovererSummary (Maybe (HashMap Text Text))

Tags associated with the resource.

discovererSummary_discovererArn :: Lens' DiscovererSummary (Maybe Text)

The ARN of the discoverer.