Amazonka.Schemas.UpdateSchema

Description

Updates the schema definition

Inactive schemas will be deleted after two years.

Creating a Request

data UpdateSchema

See: newUpdateSchema smart constructor.

Constructors

UpdateSchema'

Fields

content :: Maybe Text

The source of the schema definition.

clientTokenId :: Maybe Text

The ID of the client token.

type' :: Maybe Type

The schema type for the events schema.

description :: Maybe Text

The description of the schema.

registryName :: Text

The name of the registry.

schemaName :: Text

The name of the schema.

newUpdateSchema

Arguments

:: Text
$sel:registryName:UpdateSchema'

-> Text
$sel:schemaName:UpdateSchema'

-> UpdateSchema

Create a value of UpdateSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:content:UpdateSchema', updateSchema_content - The source of the schema definition.

$sel:clientTokenId:UpdateSchema', updateSchema_clientTokenId - The ID of the client token.

$sel:type':UpdateSchema', updateSchema_type - The schema type for the events schema.

$sel:description:UpdateSchema', updateSchema_description - The description of the schema.

$sel:registryName:UpdateSchema', updateSchema_registryName - The name of the registry.

$sel:schemaName:UpdateSchema', updateSchema_schemaName - The name of the schema.

Request Lenses

updateSchema_content :: Lens' UpdateSchema (Maybe Text)

The source of the schema definition.

updateSchema_clientTokenId :: Lens' UpdateSchema (Maybe Text)

The ID of the client token.

updateSchema_type :: Lens' UpdateSchema (Maybe Type)

The schema type for the events schema.

updateSchema_description :: Lens' UpdateSchema (Maybe Text)

The description of the schema.

updateSchema_registryName :: Lens' UpdateSchema Text

The name of the registry.

updateSchema_schemaName :: Lens' UpdateSchema Text

The name of the schema.

Destructuring the Response

data UpdateSchemaResponse

See: newUpdateSchemaResponse smart constructor.

Constructors

UpdateSchemaResponse'

Fields

schemaVersion :: Maybe Text

The version number of the schema

schemaName :: Maybe Text

The name of the schema.

schemaArn :: Maybe Text

The ARN of the schema.

type' :: Maybe Text

The type of the schema.

lastModified :: Maybe POSIX

The date and time that schema was modified.

description :: Maybe Text

The description of the schema.

versionCreatedDate :: Maybe POSIX

The date the schema version was created.

tags :: Maybe (HashMap Text Text)

httpStatus :: Int

The response's http status code.

newUpdateSchemaResponse

Arguments

:: Int
$sel:httpStatus:UpdateSchemaResponse'

-> UpdateSchemaResponse

Create a value of UpdateSchemaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersion:UpdateSchemaResponse', updateSchemaResponse_schemaVersion - The version number of the schema

$sel:schemaName:UpdateSchema', updateSchemaResponse_schemaName - The name of the schema.

$sel:schemaArn:UpdateSchemaResponse', updateSchemaResponse_schemaArn - The ARN of the schema.

$sel:type':UpdateSchema', updateSchemaResponse_type - The type of the schema.

$sel:lastModified:UpdateSchemaResponse', updateSchemaResponse_lastModified - The date and time that schema was modified.

$sel:description:UpdateSchema', updateSchemaResponse_description - The description of the schema.

$sel:versionCreatedDate:UpdateSchemaResponse', updateSchemaResponse_versionCreatedDate - The date the schema version was created.

$sel:tags:UpdateSchemaResponse', updateSchemaResponse_tags - Undocumented member.

$sel:httpStatus:UpdateSchemaResponse', updateSchemaResponse_httpStatus - The response's http status code.

Response Lenses

updateSchemaResponse_schemaVersion :: Lens' UpdateSchemaResponse (Maybe Text)

The version number of the schema

updateSchemaResponse_schemaName :: Lens' UpdateSchemaResponse (Maybe Text)

The name of the schema.

updateSchemaResponse_schemaArn :: Lens' UpdateSchemaResponse (Maybe Text)

The ARN of the schema.

updateSchemaResponse_type :: Lens' UpdateSchemaResponse (Maybe Text)

The type of the schema.

updateSchemaResponse_lastModified :: Lens' UpdateSchemaResponse (Maybe UTCTime)

The date and time that schema was modified.

updateSchemaResponse_description :: Lens' UpdateSchemaResponse (Maybe Text)

The description of the schema.

updateSchemaResponse_versionCreatedDate :: Lens' UpdateSchemaResponse (Maybe UTCTime)

The date the schema version was created.

updateSchemaResponse_tags :: Lens' UpdateSchemaResponse (Maybe (HashMap Text Text))

Undocumented member.

updateSchemaResponse_httpStatus :: Lens' UpdateSchemaResponse Int

The response's http status code.