Amazonka.SecretsManager.GetRandomPassword

Description

Generates a random password of the specified complexity. This operation is intended for use in the Lambda rotation function. Per best practice, we recommend that you specify the maximum length and include every character type that the system you are generating a password for can support.

Minimum permissions

To run this command, you must have the following permissions:

secretsmanager:GetRandomPassword

Creating a Request

data GetRandomPassword

See: newGetRandomPassword smart constructor.

Constructors

GetRandomPassword'

Fields

includeSpace :: Maybe Bool

Specifies that the generated password can include the space character. The default if you do not include this switch parameter is that the space character is not included.

excludeNumbers :: Maybe Bool

Specifies that the generated password should not include digits. The default if you do not include this switch parameter is that digits can be included.

excludeLowercase :: Maybe Bool

Specifies that the generated password should not include lowercase letters. The default if you do not include this switch parameter is that lowercase letters can be included.

excludeCharacters :: Maybe Text

A string that includes characters that should not be included in the generated password. The default is that all characters from the included sets can be used.

excludePunctuation :: Maybe Bool

Specifies that the generated password should not include punctuation characters. The default if you do not include this switch parameter is that punctuation characters can be included.

The following are the punctuation characters that can be included in the generated password if you don't explicitly exclude them with ExcludeCharacters or ExcludePunctuation:

! " # $ % & ' ( ) * + , - . / : ; < = > ? @ [ \ ] ^ _ ` { | } ~

requireEachIncludedType :: Maybe Bool

A boolean value that specifies whether the generated password must include at least one of every allowed character type. The default value is True and the operation requires at least one of every character type.

excludeUppercase :: Maybe Bool

Specifies that the generated password should not include uppercase letters. The default if you do not include this switch parameter is that uppercase letters can be included.

passwordLength :: Maybe Natural

The desired length of the generated password. The default value if you do not include this parameter is 32 characters.

newGetRandomPassword :: GetRandomPassword

Create a value of GetRandomPassword with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeSpace:GetRandomPassword', getRandomPassword_includeSpace - Specifies that the generated password can include the space character. The default if you do not include this switch parameter is that the space character is not included.

$sel:excludeNumbers:GetRandomPassword', getRandomPassword_excludeNumbers - Specifies that the generated password should not include digits. The default if you do not include this switch parameter is that digits can be included.

$sel:excludeLowercase:GetRandomPassword', getRandomPassword_excludeLowercase - Specifies that the generated password should not include lowercase letters. The default if you do not include this switch parameter is that lowercase letters can be included.

$sel:excludeCharacters:GetRandomPassword', getRandomPassword_excludeCharacters - A string that includes characters that should not be included in the generated password. The default is that all characters from the included sets can be used.

$sel:excludePunctuation:GetRandomPassword', getRandomPassword_excludePunctuation - Specifies that the generated password should not include punctuation characters. The default if you do not include this switch parameter is that punctuation characters can be included.

The following are the punctuation characters that can be included in the generated password if you don't explicitly exclude them with ExcludeCharacters or ExcludePunctuation:

! " # $ % & ' ( ) * + , - . / : ; < = > ? @ [ \ ] ^ _ ` { | } ~

$sel:requireEachIncludedType:GetRandomPassword', getRandomPassword_requireEachIncludedType - A boolean value that specifies whether the generated password must include at least one of every allowed character type. The default value is True and the operation requires at least one of every character type.

$sel:excludeUppercase:GetRandomPassword', getRandomPassword_excludeUppercase - Specifies that the generated password should not include uppercase letters. The default if you do not include this switch parameter is that uppercase letters can be included.

$sel:passwordLength:GetRandomPassword', getRandomPassword_passwordLength - The desired length of the generated password. The default value if you do not include this parameter is 32 characters.

Request Lenses

getRandomPassword_includeSpace :: Lens' GetRandomPassword (Maybe Bool)

Specifies that the generated password can include the space character. The default if you do not include this switch parameter is that the space character is not included.

getRandomPassword_excludeNumbers :: Lens' GetRandomPassword (Maybe Bool)

Specifies that the generated password should not include digits. The default if you do not include this switch parameter is that digits can be included.

getRandomPassword_excludeLowercase :: Lens' GetRandomPassword (Maybe Bool)

Specifies that the generated password should not include lowercase letters. The default if you do not include this switch parameter is that lowercase letters can be included.

getRandomPassword_excludeCharacters :: Lens' GetRandomPassword (Maybe Text)

A string that includes characters that should not be included in the generated password. The default is that all characters from the included sets can be used.

getRandomPassword_excludePunctuation :: Lens' GetRandomPassword (Maybe Bool)

Specifies that the generated password should not include punctuation characters. The default if you do not include this switch parameter is that punctuation characters can be included.

The following are the punctuation characters that can be included in the generated password if you don't explicitly exclude them with ExcludeCharacters or ExcludePunctuation:

! " # $ % & ' ( ) * + , - . / : ; < = > ? @ [ \ ] ^ _ ` { | } ~

getRandomPassword_requireEachIncludedType :: Lens' GetRandomPassword (Maybe Bool)

A boolean value that specifies whether the generated password must include at least one of every allowed character type. The default value is True and the operation requires at least one of every character type.

getRandomPassword_excludeUppercase :: Lens' GetRandomPassword (Maybe Bool)

Specifies that the generated password should not include uppercase letters. The default if you do not include this switch parameter is that uppercase letters can be included.

getRandomPassword_passwordLength :: Lens' GetRandomPassword (Maybe Natural)

The desired length of the generated password. The default value if you do not include this parameter is 32 characters.

Destructuring the Response

data GetRandomPasswordResponse

See: newGetRandomPasswordResponse smart constructor.

Constructors

GetRandomPasswordResponse'

Fields

randomPassword :: Maybe (Sensitive Text)

A string with the generated password.

httpStatus :: Int

The response's http status code.

newGetRandomPasswordResponse

Arguments

:: Int
$sel:httpStatus:GetRandomPasswordResponse'

-> GetRandomPasswordResponse

Create a value of GetRandomPasswordResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:randomPassword:GetRandomPasswordResponse', getRandomPasswordResponse_randomPassword - A string with the generated password.

$sel:httpStatus:GetRandomPasswordResponse', getRandomPasswordResponse_httpStatus - The response's http status code.

Response Lenses

getRandomPasswordResponse_randomPassword :: Lens' GetRandomPasswordResponse (Maybe Text)

A string with the generated password.

getRandomPasswordResponse_httpStatus :: Lens' GetRandomPasswordResponse Int

The response's http status code.