Amazonka.SecretsManager.RestoreSecret

Description

Cancels the scheduled deletion of a secret by removing the DeletedDate time stamp. This makes the secret accessible to query once again.

Minimum permissions

To run this command, you must have the following permissions:

secretsmanager:RestoreSecret

Related operations

To delete a secret, use DeleteSecret.

Creating a Request

data RestoreSecret

See: newRestoreSecret smart constructor.

Constructors

RestoreSecret'

Fields

secretId :: Text

Specifies the secret that you want to restore from a previously scheduled deletion. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

For an ARN, we recommend that you specify a complete ARN rather than a partial ARN.

newRestoreSecret

Arguments

:: Text
$sel:secretId:RestoreSecret'

-> RestoreSecret

Create a value of RestoreSecret with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secretId:RestoreSecret', restoreSecret_secretId - Specifies the secret that you want to restore from a previously scheduled deletion. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

For an ARN, we recommend that you specify a complete ARN rather than a partial ARN.

Request Lenses

restoreSecret_secretId :: Lens' RestoreSecret Text

Specifies the secret that you want to restore from a previously scheduled deletion. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

For an ARN, we recommend that you specify a complete ARN rather than a partial ARN.

Destructuring the Response

data RestoreSecretResponse

See: newRestoreSecretResponse smart constructor.

Constructors

RestoreSecretResponse'

Fields

arn :: Maybe Text

The ARN of the secret that was restored.

name :: Maybe Text

The friendly name of the secret that was restored.

httpStatus :: Int

The response's http status code.

newRestoreSecretResponse

Arguments

:: Int
$sel:httpStatus:RestoreSecretResponse'

-> RestoreSecretResponse

Create a value of RestoreSecretResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:RestoreSecretResponse', restoreSecretResponse_arn - The ARN of the secret that was restored.

$sel:name:RestoreSecretResponse', restoreSecretResponse_name - The friendly name of the secret that was restored.

$sel:httpStatus:RestoreSecretResponse', restoreSecretResponse_httpStatus - The response's http status code.

Response Lenses

restoreSecretResponse_arn :: Lens' RestoreSecretResponse (Maybe Text)

The ARN of the secret that was restored.

restoreSecretResponse_name :: Lens' RestoreSecretResponse (Maybe Text)

The friendly name of the secret that was restored.

restoreSecretResponse_httpStatus :: Lens' RestoreSecretResponse Int

The response's http status code.