Amazonka.SecretsManager.RotateSecret

Description

Configures and starts the asynchronous process of rotating this secret. If you include the configuration parameters, the operation sets those values for the secret and then immediately starts a rotation. If you do not include the configuration parameters, the operation starts a rotation with the values already stored in the secret. After the rotation completes, the protected service and its clients all use the new version of the secret.

This required configuration information includes the ARN of an Amazon Web Services Lambda function and optionally, the time between scheduled rotations. The Lambda rotation function creates a new version of the secret and creates or updates the credentials on the protected service to match. After testing the new credentials, the function marks the new secret with the staging label AWSCURRENT so that your clients all immediately begin to use the new version. For more information about rotating secrets and how to configure a Lambda function to rotate the secrets for your protected service, see Rotating Secrets in Amazon Web Services Secrets Manager in the Amazon Web Services Secrets Manager User Guide.

Secrets Manager schedules the next rotation when the previous one completes. Secrets Manager schedules the date by adding the rotation interval (number of days) to the actual date of the last rotation. The service chooses the hour within that 24-hour date window randomly. The minute is also chosen somewhat randomly, but weighted towards the top of the hour and influenced by a variety of factors that help distribute load.

The rotation function must end with the versions of the secret in one of two states:

The AWSPENDING and AWSCURRENT staging labels are attached to the same version of the secret, or

The AWSPENDING staging label is not attached to any version of the secret.

If the AWSPENDING staging label is present but not attached to the same version as AWSCURRENT then any later invocation of RotateSecret assumes that a previous rotation request is still in progress and returns an error.

Minimum permissions

To run this command, you must have the following permissions:

secretsmanager:RotateSecret

lambda:InvokeFunction (on the function specified in the secret's metadata)

Related operations

To list the secrets in your account, use ListSecrets.

To get the details for a version of a secret, use DescribeSecret.

To create a new version of a secret, use CreateSecret.

To attach staging labels to or remove staging labels from a version of a secret, use UpdateSecretVersionStage.

Creating a Request

data RotateSecret

See: newRotateSecret smart constructor.

Constructors

RotateSecret'

Fields

rotationRules :: Maybe RotationRulesType

A structure that defines the rotation configuration for this secret.

clientRequestToken :: Maybe Text

(Optional) Specifies a unique identifier for the new version of the secret that helps ensure idempotency.

If you use the Amazon Web Services CLI or one of the Amazon Web Services SDK to call this operation, then you can leave this parameter empty. The CLI or SDK generates a random UUID for you and includes that in the request for this parameter. If you don't use the SDK and instead generate a raw HTTP request to the Secrets Manager service endpoint, then you must generate a ClientRequestToken yourself for new versions and include that value in the request.

You only need to specify your own value if you implement your own retry logic and want to ensure that a given secret is not created twice. We recommend that you generate a UUID-type value to ensure uniqueness within the specified secret.

Secrets Manager uses this value to prevent the accidental creation of duplicate versions if there are failures and retries during the function's processing. This value becomes the VersionId of the new version.

rotationLambdaARN :: Maybe Text

(Optional) Specifies the ARN of the Lambda function that can rotate the secret.

secretId :: Text

Specifies the secret that you want to rotate. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

For an ARN, we recommend that you specify a complete ARN rather than a partial ARN.

newRotateSecret

Arguments

:: Text
$sel:secretId:RotateSecret'

-> RotateSecret

Create a value of RotateSecret with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rotationRules:RotateSecret', rotateSecret_rotationRules - A structure that defines the rotation configuration for this secret.

$sel:clientRequestToken:RotateSecret', rotateSecret_clientRequestToken - (Optional) Specifies a unique identifier for the new version of the secret that helps ensure idempotency.

If you use the Amazon Web Services CLI or one of the Amazon Web Services SDK to call this operation, then you can leave this parameter empty. The CLI or SDK generates a random UUID for you and includes that in the request for this parameter. If you don't use the SDK and instead generate a raw HTTP request to the Secrets Manager service endpoint, then you must generate a ClientRequestToken yourself for new versions and include that value in the request.

You only need to specify your own value if you implement your own retry logic and want to ensure that a given secret is not created twice. We recommend that you generate a UUID-type value to ensure uniqueness within the specified secret.

Secrets Manager uses this value to prevent the accidental creation of duplicate versions if there are failures and retries during the function's processing. This value becomes the VersionId of the new version.

$sel:rotationLambdaARN:RotateSecret', rotateSecret_rotationLambdaARN - (Optional) Specifies the ARN of the Lambda function that can rotate the secret.

$sel:secretId:RotateSecret', rotateSecret_secretId - Specifies the secret that you want to rotate. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

For an ARN, we recommend that you specify a complete ARN rather than a partial ARN.

Request Lenses

rotateSecret_rotationRules :: Lens' RotateSecret (Maybe RotationRulesType)

A structure that defines the rotation configuration for this secret.

rotateSecret_clientRequestToken :: Lens' RotateSecret (Maybe Text)

(Optional) Specifies a unique identifier for the new version of the secret that helps ensure idempotency.

If you use the Amazon Web Services CLI or one of the Amazon Web Services SDK to call this operation, then you can leave this parameter empty. The CLI or SDK generates a random UUID for you and includes that in the request for this parameter. If you don't use the SDK and instead generate a raw HTTP request to the Secrets Manager service endpoint, then you must generate a ClientRequestToken yourself for new versions and include that value in the request.

You only need to specify your own value if you implement your own retry logic and want to ensure that a given secret is not created twice. We recommend that you generate a UUID-type value to ensure uniqueness within the specified secret.

Secrets Manager uses this value to prevent the accidental creation of duplicate versions if there are failures and retries during the function's processing. This value becomes the VersionId of the new version.

rotateSecret_rotationLambdaARN :: Lens' RotateSecret (Maybe Text)

(Optional) Specifies the ARN of the Lambda function that can rotate the secret.

rotateSecret_secretId :: Lens' RotateSecret Text

Specifies the secret that you want to rotate. You can specify either the Amazon Resource Name (ARN) or the friendly name of the secret.

For an ARN, we recommend that you specify a complete ARN rather than a partial ARN.

Destructuring the Response

data RotateSecretResponse

See: newRotateSecretResponse smart constructor.

Constructors

RotateSecretResponse'

Fields

versionId :: Maybe Text

The ID of the new version of the secret created by the rotation started by this request.

arn :: Maybe Text

The ARN of the secret.

name :: Maybe Text

The friendly name of the secret.

httpStatus :: Int

The response's http status code.

newRotateSecretResponse

Arguments

:: Int
$sel:httpStatus:RotateSecretResponse'

-> RotateSecretResponse

Create a value of RotateSecretResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:RotateSecretResponse', rotateSecretResponse_versionId - The ID of the new version of the secret created by the rotation started by this request.

$sel:arn:RotateSecretResponse', rotateSecretResponse_arn - The ARN of the secret.

$sel:name:RotateSecretResponse', rotateSecretResponse_name - The friendly name of the secret.

$sel:httpStatus:RotateSecretResponse', rotateSecretResponse_httpStatus - The response's http status code.

Response Lenses

rotateSecretResponse_versionId :: Lens' RotateSecretResponse (Maybe Text)

The ID of the new version of the secret created by the rotation started by this request.

rotateSecretResponse_arn :: Lens' RotateSecretResponse (Maybe Text)

The ARN of the secret.

rotateSecretResponse_name :: Lens' RotateSecretResponse (Maybe Text)

The friendly name of the secret.

rotateSecretResponse_httpStatus :: Lens' RotateSecretResponse Int

The response's http status code.