Amazonka.SecretsManager.Types.SecretListEntry

Description

Documentation

data SecretListEntry

A structure that contains the details about a secret. It does not include the encrypted SecretString and SecretBinary values. To get those values, use the GetSecretValue operation.

See: newSecretListEntry smart constructor.

Constructors

SecretListEntry'

Fields

lastChangedDate :: Maybe POSIX

The last date and time that this secret was modified in any way.

primaryRegion :: Maybe Text

The Region where Secrets Manager originated the secret.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the secret.

For more information about ARNs in Secrets Manager, see Policy Resources in the Amazon Web Services Secrets Manager User Guide.

secretVersionsToStages :: Maybe (HashMap Text (NonEmpty Text))

A list of all of the currently assigned SecretVersionStage staging labels and the SecretVersionId attached to each one. Staging labels are used to keep track of the different versions during the rotation process.

A version that does not have any SecretVersionStage is considered deprecated and subject to deletion. Such versions are not included in this list.

rotationRules :: Maybe RotationRulesType

A structure that defines the rotation configuration for the secret.

deletedDate :: Maybe POSIX

The date and time the deletion of the secret occurred. Not present on active secrets. The secret can be recovered until the number of days in the recovery window has passed, as specified in the RecoveryWindowInDays parameter of the DeleteSecret operation.

rotationEnabled :: Maybe Bool

Indicates whether automatic, scheduled rotation is enabled for this secret.

createdDate :: Maybe POSIX

The date and time when a secret was created.

kmsKeyId :: Maybe Text

The ARN or alias of the Amazon Web Services KMS customer master key (CMK) used to encrypt the SecretString and SecretBinary fields in each version of the secret. If you don't provide a key, then Secrets Manager defaults to encrypting the secret fields with the default KMS CMK, the key named awssecretsmanager, for this account.

name :: Maybe Text

The friendly name of the secret. You can use forward slashes in the name to represent a path hierarchy. For example, /prod/databases/dbserver1 could represent the secret for a server named dbserver1 in the folder databases in the folder prod.

owningService :: Maybe Text

Returns the name of the service that created the secret.

lastRotatedDate :: Maybe POSIX

The most recent date and time that the Secrets Manager rotation process was successfully completed. This value is null if the secret hasn't ever rotated.

lastAccessedDate :: Maybe POSIX

The last date that this secret was accessed. This value is truncated to midnight of the date and therefore shows only the date, not the time.

description :: Maybe Text

The user-provided description of the secret.

rotationLambdaARN :: Maybe Text

The ARN of an Amazon Web Services Lambda function invoked by Secrets Manager to rotate and expire the secret either automatically per the schedule or manually by a call to RotateSecret.

tags :: Maybe [Tag]

The list of user-defined tags associated with the secret. To add tags to a secret, use TagResource. To remove tags, use UntagResource.

newSecretListEntry :: SecretListEntry

Create a value of SecretListEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastChangedDate:SecretListEntry', secretListEntry_lastChangedDate - The last date and time that this secret was modified in any way.

$sel:primaryRegion:SecretListEntry', secretListEntry_primaryRegion - The Region where Secrets Manager originated the secret.

$sel:arn:SecretListEntry', secretListEntry_arn - The Amazon Resource Name (ARN) of the secret.

For more information about ARNs in Secrets Manager, see Policy Resources in the Amazon Web Services Secrets Manager User Guide.

$sel:secretVersionsToStages:SecretListEntry', secretListEntry_secretVersionsToStages - A list of all of the currently assigned SecretVersionStage staging labels and the SecretVersionId attached to each one. Staging labels are used to keep track of the different versions during the rotation process.

A version that does not have any SecretVersionStage is considered deprecated and subject to deletion. Such versions are not included in this list.

$sel:rotationRules:SecretListEntry', secretListEntry_rotationRules - A structure that defines the rotation configuration for the secret.

$sel:deletedDate:SecretListEntry', secretListEntry_deletedDate - The date and time the deletion of the secret occurred. Not present on active secrets. The secret can be recovered until the number of days in the recovery window has passed, as specified in the RecoveryWindowInDays parameter of the DeleteSecret operation.

$sel:rotationEnabled:SecretListEntry', secretListEntry_rotationEnabled - Indicates whether automatic, scheduled rotation is enabled for this secret.

$sel:createdDate:SecretListEntry', secretListEntry_createdDate - The date and time when a secret was created.

$sel:kmsKeyId:SecretListEntry', secretListEntry_kmsKeyId - The ARN or alias of the Amazon Web Services KMS customer master key (CMK) used to encrypt the SecretString and SecretBinary fields in each version of the secret. If you don't provide a key, then Secrets Manager defaults to encrypting the secret fields with the default KMS CMK, the key named awssecretsmanager, for this account.

$sel:name:SecretListEntry', secretListEntry_name - The friendly name of the secret. You can use forward slashes in the name to represent a path hierarchy. For example, /prod/databases/dbserver1 could represent the secret for a server named dbserver1 in the folder databases in the folder prod.

$sel:owningService:SecretListEntry', secretListEntry_owningService - Returns the name of the service that created the secret.

$sel:lastRotatedDate:SecretListEntry', secretListEntry_lastRotatedDate - The most recent date and time that the Secrets Manager rotation process was successfully completed. This value is null if the secret hasn't ever rotated.

$sel:lastAccessedDate:SecretListEntry', secretListEntry_lastAccessedDate - The last date that this secret was accessed. This value is truncated to midnight of the date and therefore shows only the date, not the time.

$sel:description:SecretListEntry', secretListEntry_description - The user-provided description of the secret.

$sel:rotationLambdaARN:SecretListEntry', secretListEntry_rotationLambdaARN - The ARN of an Amazon Web Services Lambda function invoked by Secrets Manager to rotate and expire the secret either automatically per the schedule or manually by a call to RotateSecret.

$sel:tags:SecretListEntry', secretListEntry_tags - The list of user-defined tags associated with the secret. To add tags to a secret, use TagResource. To remove tags, use UntagResource.

secretListEntry_lastChangedDate :: Lens' SecretListEntry (Maybe UTCTime)

The last date and time that this secret was modified in any way.

secretListEntry_primaryRegion :: Lens' SecretListEntry (Maybe Text)

The Region where Secrets Manager originated the secret.

secretListEntry_arn :: Lens' SecretListEntry (Maybe Text)

The Amazon Resource Name (ARN) of the secret.

For more information about ARNs in Secrets Manager, see Policy Resources in the Amazon Web Services Secrets Manager User Guide.

secretListEntry_secretVersionsToStages :: Lens' SecretListEntry (Maybe (HashMap Text (NonEmpty Text)))

A list of all of the currently assigned SecretVersionStage staging labels and the SecretVersionId attached to each one. Staging labels are used to keep track of the different versions during the rotation process.

A version that does not have any SecretVersionStage is considered deprecated and subject to deletion. Such versions are not included in this list.

secretListEntry_rotationRules :: Lens' SecretListEntry (Maybe RotationRulesType)

A structure that defines the rotation configuration for the secret.

secretListEntry_deletedDate :: Lens' SecretListEntry (Maybe UTCTime)

The date and time the deletion of the secret occurred. Not present on active secrets. The secret can be recovered until the number of days in the recovery window has passed, as specified in the RecoveryWindowInDays parameter of the DeleteSecret operation.

secretListEntry_rotationEnabled :: Lens' SecretListEntry (Maybe Bool)

Indicates whether automatic, scheduled rotation is enabled for this secret.

secretListEntry_createdDate :: Lens' SecretListEntry (Maybe UTCTime)

The date and time when a secret was created.

secretListEntry_kmsKeyId :: Lens' SecretListEntry (Maybe Text)

The ARN or alias of the Amazon Web Services KMS customer master key (CMK) used to encrypt the SecretString and SecretBinary fields in each version of the secret. If you don't provide a key, then Secrets Manager defaults to encrypting the secret fields with the default KMS CMK, the key named awssecretsmanager, for this account.

secretListEntry_name :: Lens' SecretListEntry (Maybe Text)

The friendly name of the secret. You can use forward slashes in the name to represent a path hierarchy. For example, /prod/databases/dbserver1 could represent the secret for a server named dbserver1 in the folder databases in the folder prod.

secretListEntry_owningService :: Lens' SecretListEntry (Maybe Text)

Returns the name of the service that created the secret.

secretListEntry_lastRotatedDate :: Lens' SecretListEntry (Maybe UTCTime)

The most recent date and time that the Secrets Manager rotation process was successfully completed. This value is null if the secret hasn't ever rotated.

secretListEntry_lastAccessedDate :: Lens' SecretListEntry (Maybe UTCTime)

The last date that this secret was accessed. This value is truncated to midnight of the date and therefore shows only the date, not the time.

secretListEntry_description :: Lens' SecretListEntry (Maybe Text)

The user-provided description of the secret.

secretListEntry_rotationLambdaARN :: Lens' SecretListEntry (Maybe Text)

The ARN of an Amazon Web Services Lambda function invoked by Secrets Manager to rotate and expire the secret either automatically per the schedule or manually by a call to RotateSecret.

secretListEntry_tags :: Lens' SecretListEntry (Maybe [Tag])

The list of user-defined tags associated with the secret. To add tags to a secret, use TagResource. To remove tags, use UntagResource.