Amazonka.SecurityHub.BatchImportFindings

Description

Imports security findings generated from an integrated product into Security Hub. This action is requested by the integrated product to import its findings into Security Hub.

The maximum allowed size for a finding is 240 Kb. An error is returned for any finding larger than 240 Kb.

After a finding is created, BatchImportFindings cannot be used to update the following finding fields and objects, which Security Hub customers use to manage their investigation workflow.

Note

UserDefinedFields

VerificationState

Workflow

Finding providers also should not use BatchImportFindings to update the following attributes.

Confidence

Criticality

RelatedFindings

Severity

Types

Instead, finding providers use FindingProviderFields to provide values for these attributes.

Creating a Request

data BatchImportFindings

See: newBatchImportFindings smart constructor.

Constructors

BatchImportFindings'

Fields

findings :: NonEmpty AwsSecurityFinding

A list of findings to import. To successfully import a finding, it must follow the Amazon Web Services Security Finding Format. Maximum of 100 findings per request.

newBatchImportFindings

Arguments

:: NonEmpty AwsSecurityFinding
$sel:findings:BatchImportFindings'

-> BatchImportFindings

Create a value of BatchImportFindings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:findings:BatchImportFindings', batchImportFindings_findings - A list of findings to import. To successfully import a finding, it must follow the Amazon Web Services Security Finding Format. Maximum of 100 findings per request.

Request Lenses

batchImportFindings_findings :: Lens' BatchImportFindings (NonEmpty AwsSecurityFinding)

A list of findings to import. To successfully import a finding, it must follow the Amazon Web Services Security Finding Format. Maximum of 100 findings per request.

Destructuring the Response

data BatchImportFindingsResponse

See: newBatchImportFindingsResponse smart constructor.

Constructors

BatchImportFindingsResponse'

Fields

failedFindings :: Maybe [ImportFindingsError]

The list of findings that failed to import.

httpStatus :: Int

The response's http status code.

failedCount :: Int

The number of findings that failed to import.

successCount :: Int

The number of findings that were successfully imported.

newBatchImportFindingsResponse

Arguments

:: Int
$sel:httpStatus:BatchImportFindingsResponse'

-> Int
$sel:failedCount:BatchImportFindingsResponse'

-> Int
$sel:successCount:BatchImportFindingsResponse'

-> BatchImportFindingsResponse

Create a value of BatchImportFindingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedFindings:BatchImportFindingsResponse', batchImportFindingsResponse_failedFindings - The list of findings that failed to import.

$sel:httpStatus:BatchImportFindingsResponse', batchImportFindingsResponse_httpStatus - The response's http status code.

$sel:failedCount:BatchImportFindingsResponse', batchImportFindingsResponse_failedCount - The number of findings that failed to import.

$sel:successCount:BatchImportFindingsResponse', batchImportFindingsResponse_successCount - The number of findings that were successfully imported.

Response Lenses

batchImportFindingsResponse_failedFindings :: Lens' BatchImportFindingsResponse (Maybe [ImportFindingsError])

The list of findings that failed to import.

batchImportFindingsResponse_httpStatus :: Lens' BatchImportFindingsResponse Int

The response's http status code.

batchImportFindingsResponse_failedCount :: Lens' BatchImportFindingsResponse Int

The number of findings that failed to import.

batchImportFindingsResponse_successCount :: Lens' BatchImportFindingsResponse Int

The number of findings that were successfully imported.