Amazonka.SecurityHub.CreateInsight

Description

Creates a custom insight in Security Hub. An insight is a consolidation of findings that relate to a security issue that requires attention or remediation.

To group the related findings in the insight, use the GroupByAttribute.

Creating a Request

data CreateInsight

See: newCreateInsight smart constructor.

Constructors

CreateInsight'

Fields

name :: Text

The name of the custom insight to create.

filters :: AwsSecurityFindingFilters

One or more attributes used to filter the findings included in the insight. The insight only includes findings that match the criteria defined in the filters.

groupByAttribute :: Text

The attribute used to group the findings for the insight. The grouping attribute identifies the type of item that the insight applies to. For example, if an insight is grouped by resource identifier, then the insight produces a list of resource identifiers.

newCreateInsight

Arguments

:: Text
$sel:name:CreateInsight'

-> AwsSecurityFindingFilters
$sel:filters:CreateInsight'

-> Text
$sel:groupByAttribute:CreateInsight'

-> CreateInsight

Create a value of CreateInsight with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateInsight', createInsight_name - The name of the custom insight to create.

$sel:filters:CreateInsight', createInsight_filters - One or more attributes used to filter the findings included in the insight. The insight only includes findings that match the criteria defined in the filters.

$sel:groupByAttribute:CreateInsight', createInsight_groupByAttribute - The attribute used to group the findings for the insight. The grouping attribute identifies the type of item that the insight applies to. For example, if an insight is grouped by resource identifier, then the insight produces a list of resource identifiers.

Request Lenses

createInsight_name :: Lens' CreateInsight Text

The name of the custom insight to create.

createInsight_filters :: Lens' CreateInsight AwsSecurityFindingFilters

One or more attributes used to filter the findings included in the insight. The insight only includes findings that match the criteria defined in the filters.

createInsight_groupByAttribute :: Lens' CreateInsight Text

The attribute used to group the findings for the insight. The grouping attribute identifies the type of item that the insight applies to. For example, if an insight is grouped by resource identifier, then the insight produces a list of resource identifiers.

Destructuring the Response

data CreateInsightResponse

See: newCreateInsightResponse smart constructor.

Constructors

CreateInsightResponse'

Fields

httpStatus :: Int

The response's http status code.

insightArn :: Text

The ARN of the insight created.

newCreateInsightResponse

Arguments

:: Int
$sel:httpStatus:CreateInsightResponse'

-> Text
$sel:insightArn:CreateInsightResponse'

-> CreateInsightResponse

Create a value of CreateInsightResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateInsightResponse', createInsightResponse_httpStatus - The response's http status code.

$sel:insightArn:CreateInsightResponse', createInsightResponse_insightArn - The ARN of the insight created.

Response Lenses

createInsightResponse_httpStatus :: Lens' CreateInsightResponse Int

The response's http status code.

createInsightResponse_insightArn :: Lens' CreateInsightResponse Text

The ARN of the insight created.