Amazonka.SecurityHub.DescribeHub

Description

Returns details about the Hub resource in your account, including the HubArn and the time when you enabled Security Hub.

Creating a Request

data DescribeHub

See: newDescribeHub smart constructor.

Constructors

DescribeHub'

Fields

hubArn :: Maybe Text

The ARN of the Hub resource to retrieve.

newDescribeHub :: DescribeHub

Create a value of DescribeHub with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hubArn:DescribeHub', describeHub_hubArn - The ARN of the Hub resource to retrieve.

Request Lenses

describeHub_hubArn :: Lens' DescribeHub (Maybe Text)

The ARN of the Hub resource to retrieve.

Destructuring the Response

data DescribeHubResponse

See: newDescribeHubResponse smart constructor.

Constructors

DescribeHubResponse'

Fields

autoEnableControls :: Maybe Bool

Whether to automatically enable new controls when they are added to standards that are enabled.

If set to true, then new controls for enabled standards are enabled automatically. If set to false, then new controls are not enabled.

subscribedAt :: Maybe Text

The date and time when Security Hub was enabled in the account.

hubArn :: Maybe Text

The ARN of the Hub resource that was retrieved.

httpStatus :: Int

The response's http status code.

newDescribeHubResponse

Arguments

:: Int
$sel:httpStatus:DescribeHubResponse'

-> DescribeHubResponse

Create a value of DescribeHubResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoEnableControls:DescribeHubResponse', describeHubResponse_autoEnableControls - Whether to automatically enable new controls when they are added to standards that are enabled.

If set to true, then new controls for enabled standards are enabled automatically. If set to false, then new controls are not enabled.

$sel:subscribedAt:DescribeHubResponse', describeHubResponse_subscribedAt - The date and time when Security Hub was enabled in the account.

$sel:hubArn:DescribeHub', describeHubResponse_hubArn - The ARN of the Hub resource that was retrieved.

$sel:httpStatus:DescribeHubResponse', describeHubResponse_httpStatus - The response's http status code.

Response Lenses

describeHubResponse_autoEnableControls :: Lens' DescribeHubResponse (Maybe Bool)

Whether to automatically enable new controls when they are added to standards that are enabled.

If set to true, then new controls for enabled standards are enabled automatically. If set to false, then new controls are not enabled.

describeHubResponse_subscribedAt :: Lens' DescribeHubResponse (Maybe Text)

The date and time when Security Hub was enabled in the account.

describeHubResponse_hubArn :: Lens' DescribeHubResponse (Maybe Text)

The ARN of the Hub resource that was retrieved.

describeHubResponse_httpStatus :: Lens' DescribeHubResponse Int

The response's http status code.