Amazonka.SecurityHub.DescribeStandardsControls

Description

Returns a list of security standards controls.

For each control, the results include information about whether it is currently enabled, the severity, and a link to remediation information.

This operation returns paginated results.

Creating a Request

data DescribeStandardsControls

See: newDescribeStandardsControls smart constructor.

Constructors

DescribeStandardsControls'

Fields

nextToken :: Maybe Text

The token that is required for pagination. On your first call to the DescribeStandardsControls operation, set the value of this parameter to NULL.

For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

maxResults :: Maybe Natural

The maximum number of security standard controls to return.

standardsSubscriptionArn :: Text

The ARN of a resource that represents your subscription to a supported standard. To get the subscription ARNs of the standards you have enabled, use the GetEnabledStandards operation.

newDescribeStandardsControls

Arguments

:: Text
$sel:standardsSubscriptionArn:DescribeStandardsControls'

-> DescribeStandardsControls

Create a value of DescribeStandardsControls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeStandardsControls', describeStandardsControls_nextToken - The token that is required for pagination. On your first call to the DescribeStandardsControls operation, set the value of this parameter to NULL.

For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

$sel:maxResults:DescribeStandardsControls', describeStandardsControls_maxResults - The maximum number of security standard controls to return.

$sel:standardsSubscriptionArn:DescribeStandardsControls', describeStandardsControls_standardsSubscriptionArn - The ARN of a resource that represents your subscription to a supported standard. To get the subscription ARNs of the standards you have enabled, use the GetEnabledStandards operation.

Request Lenses

describeStandardsControls_nextToken :: Lens' DescribeStandardsControls (Maybe Text)

The token that is required for pagination. On your first call to the DescribeStandardsControls operation, set the value of this parameter to NULL.

For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

describeStandardsControls_maxResults :: Lens' DescribeStandardsControls (Maybe Natural)

The maximum number of security standard controls to return.

describeStandardsControls_standardsSubscriptionArn :: Lens' DescribeStandardsControls Text

The ARN of a resource that represents your subscription to a supported standard. To get the subscription ARNs of the standards you have enabled, use the GetEnabledStandards operation.

Destructuring the Response

data DescribeStandardsControlsResponse

See: newDescribeStandardsControlsResponse smart constructor.

Constructors

DescribeStandardsControlsResponse'

Fields

controls :: Maybe [StandardsControl]

A list of security standards controls.

nextToken :: Maybe Text

The pagination token to use to request the next page of results.

httpStatus :: Int

The response's http status code.

newDescribeStandardsControlsResponse

Arguments

:: Int
$sel:httpStatus:DescribeStandardsControlsResponse'

-> DescribeStandardsControlsResponse

Create a value of DescribeStandardsControlsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:controls:DescribeStandardsControlsResponse', describeStandardsControlsResponse_controls - A list of security standards controls.

$sel:nextToken:DescribeStandardsControls', describeStandardsControlsResponse_nextToken - The pagination token to use to request the next page of results.

$sel:httpStatus:DescribeStandardsControlsResponse', describeStandardsControlsResponse_httpStatus - The response's http status code.

Response Lenses

describeStandardsControlsResponse_controls :: Lens' DescribeStandardsControlsResponse (Maybe [StandardsControl])

A list of security standards controls.

describeStandardsControlsResponse_nextToken :: Lens' DescribeStandardsControlsResponse (Maybe Text)

The pagination token to use to request the next page of results.

describeStandardsControlsResponse_httpStatus :: Lens' DescribeStandardsControlsResponse Int

The response's http status code.