Amazonka.SecurityHub.GetAdministratorAccount

Description

Provides the details for the Security Hub administrator account for the current member account.

Can be used by both member accounts that are managed using Organizations and accounts that were invited manually.

Creating a Request

data GetAdministratorAccount

See: newGetAdministratorAccount smart constructor.

Constructors

GetAdministratorAccount'

newGetAdministratorAccount :: GetAdministratorAccount

Create a value of GetAdministratorAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetAdministratorAccountResponse

See: newGetAdministratorAccountResponse smart constructor.

Constructors

GetAdministratorAccountResponse'

Fields

administrator :: Maybe Invitation

httpStatus :: Int

The response's http status code.

newGetAdministratorAccountResponse

Arguments

:: Int
$sel:httpStatus:GetAdministratorAccountResponse'

-> GetAdministratorAccountResponse

Create a value of GetAdministratorAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:administrator:GetAdministratorAccountResponse', getAdministratorAccountResponse_administrator - Undocumented member.

$sel:httpStatus:GetAdministratorAccountResponse', getAdministratorAccountResponse_httpStatus - The response's http status code.

Response Lenses

getAdministratorAccountResponse_administrator :: Lens' GetAdministratorAccountResponse (Maybe Invitation)

Undocumented member.

getAdministratorAccountResponse_httpStatus :: Lens' GetAdministratorAccountResponse Int

The response's http status code.