Amazonka.SecurityHub.Types.AwsApiCallAction

Description

Documentation

data AwsApiCallAction

Provided if ActionType is AWS_API_CALL. It provides details about the API call that was detected.

See: newAwsApiCallAction smart constructor.

Constructors

AwsApiCallAction'

Fields

remoteIpDetails :: Maybe ActionRemoteIpDetails

Provided if CallerType is remoteIp. Provides information about the remote IP address that the API call originated from.

firstSeen :: Maybe Text

An ISO8601-formatted timestamp that indicates when the API call was first observed.

callerType :: Maybe Text

Indicates whether the API call originated from a remote IP address (remoteip) or from a DNS domain (domain).

affectedResources :: Maybe (HashMap Text Text)

Identifies the resources that were affected by the API call.

lastSeen :: Maybe Text

An ISO8601-formatted timestamp that indicates when the API call was most recently observed.

domainDetails :: Maybe AwsApiCallActionDomainDetails

Provided if CallerType is domain. Provides information about the DNS domain that the API call originated from.

serviceName :: Maybe Text

The name of the Amazon Web Services service that the API method belongs to.

api :: Maybe Text

The name of the API method that was issued.

newAwsApiCallAction :: AwsApiCallAction

Create a value of AwsApiCallAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteIpDetails:AwsApiCallAction', awsApiCallAction_remoteIpDetails - Provided if CallerType is remoteIp. Provides information about the remote IP address that the API call originated from.

$sel:firstSeen:AwsApiCallAction', awsApiCallAction_firstSeen - An ISO8601-formatted timestamp that indicates when the API call was first observed.

$sel:callerType:AwsApiCallAction', awsApiCallAction_callerType - Indicates whether the API call originated from a remote IP address (remoteip) or from a DNS domain (domain).

$sel:affectedResources:AwsApiCallAction', awsApiCallAction_affectedResources - Identifies the resources that were affected by the API call.

$sel:lastSeen:AwsApiCallAction', awsApiCallAction_lastSeen - An ISO8601-formatted timestamp that indicates when the API call was most recently observed.

$sel:domainDetails:AwsApiCallAction', awsApiCallAction_domainDetails - Provided if CallerType is domain. Provides information about the DNS domain that the API call originated from.

$sel:serviceName:AwsApiCallAction', awsApiCallAction_serviceName - The name of the Amazon Web Services service that the API method belongs to.

$sel:api:AwsApiCallAction', awsApiCallAction_api - The name of the API method that was issued.

awsApiCallAction_remoteIpDetails :: Lens' AwsApiCallAction (Maybe ActionRemoteIpDetails)

Provided if CallerType is remoteIp. Provides information about the remote IP address that the API call originated from.

awsApiCallAction_firstSeen :: Lens' AwsApiCallAction (Maybe Text)

An ISO8601-formatted timestamp that indicates when the API call was first observed.

awsApiCallAction_callerType :: Lens' AwsApiCallAction (Maybe Text)

Indicates whether the API call originated from a remote IP address (remoteip) or from a DNS domain (domain).

awsApiCallAction_affectedResources :: Lens' AwsApiCallAction (Maybe (HashMap Text Text))

Identifies the resources that were affected by the API call.

awsApiCallAction_lastSeen :: Lens' AwsApiCallAction (Maybe Text)

An ISO8601-formatted timestamp that indicates when the API call was most recently observed.

awsApiCallAction_domainDetails :: Lens' AwsApiCallAction (Maybe AwsApiCallActionDomainDetails)

Provided if CallerType is domain. Provides information about the DNS domain that the API call originated from.

awsApiCallAction_serviceName :: Lens' AwsApiCallAction (Maybe Text)

The name of the Amazon Web Services service that the API method belongs to.

awsApiCallAction_api :: Lens' AwsApiCallAction (Maybe Text)

The name of the API method that was issued.