Amazonka.SecurityHub.Types.AwsApiGatewayRestApiDetails

Description

Documentation

data AwsApiGatewayRestApiDetails

Contains information about a REST API in version 1 of Amazon API Gateway.

See: newAwsApiGatewayRestApiDetails smart constructor.

Constructors

AwsApiGatewayRestApiDetails'

Fields

minimumCompressionSize :: Maybe Int

The minimum size in bytes of a payload before compression is enabled.

If null, then compression is disabled.

If 0, then all payloads are compressed.

binaryMediaTypes :: Maybe [Text]

The list of binary media types supported by the REST API.

createdDate :: Maybe Text

Indicates when the API was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

name :: Maybe Text

The name of the REST API.

apiKeySource :: Maybe Text

The source of the API key for metering requests according to a usage plan.

HEADER indicates whether to read the API key from the X-API-Key header of a request.

AUTHORIZER indicates whether to read the API key from the UsageIdentifierKey from a custom authorizer.

version :: Maybe Text

The version identifier for the REST API.

id :: Maybe Text

The identifier of the REST API.

endpointConfiguration :: Maybe AwsApiGatewayEndpointConfiguration

The endpoint configuration of the REST API.

description :: Maybe Text

A description of the REST API.

newAwsApiGatewayRestApiDetails :: AwsApiGatewayRestApiDetails

Create a value of AwsApiGatewayRestApiDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minimumCompressionSize:AwsApiGatewayRestApiDetails', awsApiGatewayRestApiDetails_minimumCompressionSize - The minimum size in bytes of a payload before compression is enabled.

If null, then compression is disabled.

If 0, then all payloads are compressed.

$sel:binaryMediaTypes:AwsApiGatewayRestApiDetails', awsApiGatewayRestApiDetails_binaryMediaTypes - The list of binary media types supported by the REST API.

$sel:createdDate:AwsApiGatewayRestApiDetails', awsApiGatewayRestApiDetails_createdDate - Indicates when the API was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:name:AwsApiGatewayRestApiDetails', awsApiGatewayRestApiDetails_name - The name of the REST API.

$sel:apiKeySource:AwsApiGatewayRestApiDetails', awsApiGatewayRestApiDetails_apiKeySource - The source of the API key for metering requests according to a usage plan.

HEADER indicates whether to read the API key from the X-API-Key header of a request.

AUTHORIZER indicates whether to read the API key from the UsageIdentifierKey from a custom authorizer.

$sel:version:AwsApiGatewayRestApiDetails', awsApiGatewayRestApiDetails_version - The version identifier for the REST API.

$sel:id:AwsApiGatewayRestApiDetails', awsApiGatewayRestApiDetails_id - The identifier of the REST API.

$sel:endpointConfiguration:AwsApiGatewayRestApiDetails', awsApiGatewayRestApiDetails_endpointConfiguration - The endpoint configuration of the REST API.

$sel:description:AwsApiGatewayRestApiDetails', awsApiGatewayRestApiDetails_description - A description of the REST API.

awsApiGatewayRestApiDetails_minimumCompressionSize :: Lens' AwsApiGatewayRestApiDetails (Maybe Int)

The minimum size in bytes of a payload before compression is enabled.

If null, then compression is disabled.

If 0, then all payloads are compressed.

awsApiGatewayRestApiDetails_binaryMediaTypes :: Lens' AwsApiGatewayRestApiDetails (Maybe [Text])

The list of binary media types supported by the REST API.

awsApiGatewayRestApiDetails_createdDate :: Lens' AwsApiGatewayRestApiDetails (Maybe Text)

Indicates when the API was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsApiGatewayRestApiDetails_name :: Lens' AwsApiGatewayRestApiDetails (Maybe Text)

The name of the REST API.

awsApiGatewayRestApiDetails_apiKeySource :: Lens' AwsApiGatewayRestApiDetails (Maybe Text)

The source of the API key for metering requests according to a usage plan.

HEADER indicates whether to read the API key from the X-API-Key header of a request.

AUTHORIZER indicates whether to read the API key from the UsageIdentifierKey from a custom authorizer.

awsApiGatewayRestApiDetails_version :: Lens' AwsApiGatewayRestApiDetails (Maybe Text)

The version identifier for the REST API.

awsApiGatewayRestApiDetails_id :: Lens' AwsApiGatewayRestApiDetails (Maybe Text)

The identifier of the REST API.

awsApiGatewayRestApiDetails_endpointConfiguration :: Lens' AwsApiGatewayRestApiDetails (Maybe AwsApiGatewayEndpointConfiguration)

The endpoint configuration of the REST API.

awsApiGatewayRestApiDetails_description :: Lens' AwsApiGatewayRestApiDetails (Maybe Text)

A description of the REST API.