Amazonka.SecurityHub.Types.AwsAutoScalingAutoScalingGroupDetails

Description

Documentation

data AwsAutoScalingAutoScalingGroupDetails

Provides details about an auto scaling group.

See: newAwsAutoScalingAutoScalingGroupDetails smart constructor.

Constructors

AwsAutoScalingAutoScalingGroupDetails'

Fields

createdTime :: Maybe Text

Indicates when the auto scaling group was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

healthCheckGracePeriod :: Maybe Int

The amount of time, in seconds, that Amazon EC2 Auto Scaling waits before it checks the health status of an EC2 instance that has come into service.

launchConfigurationName :: Maybe Text

The name of the launch configuration.

healthCheckType :: Maybe Text

The service to use for the health checks.

loadBalancerNames :: Maybe [Text]

The list of load balancers associated with the group.

newAwsAutoScalingAutoScalingGroupDetails :: AwsAutoScalingAutoScalingGroupDetails

Create a value of AwsAutoScalingAutoScalingGroupDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdTime:AwsAutoScalingAutoScalingGroupDetails', awsAutoScalingAutoScalingGroupDetails_createdTime - Indicates when the auto scaling group was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:healthCheckGracePeriod:AwsAutoScalingAutoScalingGroupDetails', awsAutoScalingAutoScalingGroupDetails_healthCheckGracePeriod - The amount of time, in seconds, that Amazon EC2 Auto Scaling waits before it checks the health status of an EC2 instance that has come into service.

$sel:launchConfigurationName:AwsAutoScalingAutoScalingGroupDetails', awsAutoScalingAutoScalingGroupDetails_launchConfigurationName - The name of the launch configuration.

$sel:healthCheckType:AwsAutoScalingAutoScalingGroupDetails', awsAutoScalingAutoScalingGroupDetails_healthCheckType - The service to use for the health checks.

$sel:loadBalancerNames:AwsAutoScalingAutoScalingGroupDetails', awsAutoScalingAutoScalingGroupDetails_loadBalancerNames - The list of load balancers associated with the group.

awsAutoScalingAutoScalingGroupDetails_createdTime :: Lens' AwsAutoScalingAutoScalingGroupDetails (Maybe Text)

Indicates when the auto scaling group was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsAutoScalingAutoScalingGroupDetails_healthCheckGracePeriod :: Lens' AwsAutoScalingAutoScalingGroupDetails (Maybe Int)

The amount of time, in seconds, that Amazon EC2 Auto Scaling waits before it checks the health status of an EC2 instance that has come into service.

awsAutoScalingAutoScalingGroupDetails_launchConfigurationName :: Lens' AwsAutoScalingAutoScalingGroupDetails (Maybe Text)

The name of the launch configuration.

awsAutoScalingAutoScalingGroupDetails_healthCheckType :: Lens' AwsAutoScalingAutoScalingGroupDetails (Maybe Text)

The service to use for the health checks.

awsAutoScalingAutoScalingGroupDetails_loadBalancerNames :: Lens' AwsAutoScalingAutoScalingGroupDetails (Maybe [Text])

The list of load balancers associated with the group.