Amazonka.SecurityHub.Types.AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails

Description

Documentation

data AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails

A block device for the instance.

See: newAwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails smart constructor.

Constructors

AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails'

Fields

virtualName :: Maybe Text

The name of the virtual device (for example, ephemeral0).

You can provide either VirtualName or Ebs, but not both.

noDevice :: Maybe Bool

Whether to suppress the device that is included in the block device mapping of the Amazon Machine Image (AMI).

If NoDevice is true, then you cannot specify Ebs.>

ebs :: Maybe AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails

Parameters that are used to automatically set up Amazon EBS volumes when an instance is launched.

deviceName :: Maybe Text

The device name that is exposed to the EC2 instance. For example, /dev/sdh or xvdh.

newAwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails :: AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails

Create a value of AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualName:AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails', awsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails_virtualName - The name of the virtual device (for example, ephemeral0).

You can provide either VirtualName or Ebs, but not both.

$sel:noDevice:AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails', awsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails_noDevice - Whether to suppress the device that is included in the block device mapping of the Amazon Machine Image (AMI).

If NoDevice is true, then you cannot specify Ebs.>

$sel:ebs:AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails', awsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails_ebs - Parameters that are used to automatically set up Amazon EBS volumes when an instance is launched.

$sel:deviceName:AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails', awsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails_deviceName - The device name that is exposed to the EC2 instance. For example, /dev/sdh or xvdh.

awsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails_virtualName :: Lens' AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails (Maybe Text)

The name of the virtual device (for example, ephemeral0).

You can provide either VirtualName or Ebs, but not both.

awsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails_noDevice :: Lens' AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails (Maybe Bool)

Whether to suppress the device that is included in the block device mapping of the Amazon Machine Image (AMI).

If NoDevice is true, then you cannot specify Ebs.>

awsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails_ebs :: Lens' AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails (Maybe AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails)

Parameters that are used to automatically set up Amazon EBS volumes when an instance is launched.

awsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails_deviceName :: Lens' AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails (Maybe Text)

The device name that is exposed to the EC2 instance. For example, /dev/sdh or xvdh.