Amazonka.SecurityHub.Types.AwsCertificateManagerCertificateDetails

Description

Documentation

data AwsCertificateManagerCertificateDetails

Provides details about an Certificate Manager certificate.

See: newAwsCertificateManagerCertificateDetails smart constructor.

Constructors

AwsCertificateManagerCertificateDetails'

Fields

subject :: Maybe Text

The name of the entity that is associated with the public key contained in the certificate.

status :: Maybe Text

The status of the certificate.

Valid values: PENDING_VALIDATION | ISSUED | INACTIVE | EXPIRED | VALIDATION_TIMED_OUT | REVOKED | FAILED

failureReason :: Maybe Text

For a failed certificate request, the reason for the failure.

Valid values: NO_AVAILABLE_CONTACTS | ADDITIONAL_VERIFICATION_REQUIRED | DOMAIN_NOT_ALLOWED | INVALID_PUBLIC_DOMAIN | DOMAIN_VALIDATION_DENIED | CAA_ERROR | PCA_LIMIT_EXCEEDED | PCA_INVALID_ARN | PCA_INVALID_STATE | PCA_REQUEST_FAILED | PCA_NAME_CONSTRAINTS_VALIDATION | PCA_RESOURCE_NOT_FOUND | PCA_INVALID_ARGS | PCA_INVALID_DURATION | PCA_ACCESS_DENIED | SLR_NOT_FOUND | OTHER

subjectAlternativeNames :: Maybe [Text]

One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate.

The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

inUseBy :: Maybe [Text]

The list of ARNs for the Amazon Web Services resources that use the certificate.

createdAt :: Maybe Text

Indicates when the certificate was requested.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

serial :: Maybe Text

The serial number of the certificate.

renewalEligibility :: Maybe Text

Whether the certificate is eligible for renewal.

Valid values: ELIGIBLE | INELIGIBLE

extendedKeyUsages :: Maybe [AwsCertificateManagerCertificateExtendedKeyUsage]

Contains a list of Extended Key Usage X.509 v3 extension objects. Each object specifies a purpose for which the certificate public key can be used and consists of a name and an object identifier (OID).

importedAt :: Maybe Text

Indicates when the certificate was imported. Provided if the certificate type is IMPORTED.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

keyUsages :: Maybe [AwsCertificateManagerCertificateKeyUsage]

A list of key usage X.509 v3 extension objects.

notBefore :: Maybe Text

The time before which the certificate is not valid.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

domainName :: Maybe Text

The fully qualified domain name (FQDN), such as www.example.com, that is secured by the certificate.

renewalSummary :: Maybe AwsCertificateManagerCertificateRenewalSummary

Information about the status of the Certificate Manager managed renewal for the certificate. Provided only when the certificate type is AMAZON_ISSUED.

keyAlgorithm :: Maybe Text

The algorithm that was used to generate the public-private key pair.

Valid values: RSA_2048 | RSA_1024 | RSA_4096 | EC_prime256v1 | EC_secp384r1 | EC_secp521r1

type' :: Maybe Text

The source of the certificate. For certificates that Certificate Manager provides, Type is AMAZON_ISSUED. For certificates that are imported with ImportCertificate, Type is IMPORTED.

Valid values: IMPORTED | AMAZON_ISSUED | PRIVATE

options :: Maybe AwsCertificateManagerCertificateOptions

Provides a value that specifies whether to add the certificate to a transparency log.

issuedAt :: Maybe Text

Indicates when the certificate was issued. Provided if the certificate type is AMAZON_ISSUED.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

signatureAlgorithm :: Maybe Text

The algorithm that was used to sign the certificate.

domainValidationOptions :: Maybe [AwsCertificateManagerCertificateDomainValidationOption]

Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request.

Only provided if the certificate type is AMAZON_ISSUED.

issuer :: Maybe Text

The name of the certificate authority that issued and signed the certificate.

notAfter :: Maybe Text

The time after which the certificate becomes invalid.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

certificateAuthorityArn :: Maybe Text

The ARN of the private certificate authority (CA) that will be used to issue the certificate.

newAwsCertificateManagerCertificateDetails :: AwsCertificateManagerCertificateDetails

Create a value of AwsCertificateManagerCertificateDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subject:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_subject - The name of the entity that is associated with the public key contained in the certificate.

$sel:status:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_status - The status of the certificate.

Valid values: PENDING_VALIDATION | ISSUED | INACTIVE | EXPIRED | VALIDATION_TIMED_OUT | REVOKED | FAILED

$sel:failureReason:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_failureReason - For a failed certificate request, the reason for the failure.

Valid values: NO_AVAILABLE_CONTACTS | ADDITIONAL_VERIFICATION_REQUIRED | DOMAIN_NOT_ALLOWED | INVALID_PUBLIC_DOMAIN | DOMAIN_VALIDATION_DENIED | CAA_ERROR | PCA_LIMIT_EXCEEDED | PCA_INVALID_ARN | PCA_INVALID_STATE | PCA_REQUEST_FAILED | PCA_NAME_CONSTRAINTS_VALIDATION | PCA_RESOURCE_NOT_FOUND | PCA_INVALID_ARGS | PCA_INVALID_DURATION | PCA_ACCESS_DENIED | SLR_NOT_FOUND | OTHER

$sel:subjectAlternativeNames:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_subjectAlternativeNames - One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate.

The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

$sel:inUseBy:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_inUseBy - The list of ARNs for the Amazon Web Services resources that use the certificate.

$sel:createdAt:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_createdAt - Indicates when the certificate was requested.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:serial:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_serial - The serial number of the certificate.

$sel:renewalEligibility:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_renewalEligibility - Whether the certificate is eligible for renewal.

Valid values: ELIGIBLE | INELIGIBLE

$sel:extendedKeyUsages:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_extendedKeyUsages - Contains a list of Extended Key Usage X.509 v3 extension objects. Each object specifies a purpose for which the certificate public key can be used and consists of a name and an object identifier (OID).

$sel:importedAt:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_importedAt - Indicates when the certificate was imported. Provided if the certificate type is IMPORTED.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:keyUsages:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_keyUsages - A list of key usage X.509 v3 extension objects.

$sel:notBefore:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_notBefore - The time before which the certificate is not valid.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:domainName:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_domainName - The fully qualified domain name (FQDN), such as www.example.com, that is secured by the certificate.

$sel:renewalSummary:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_renewalSummary - Information about the status of the Certificate Manager managed renewal for the certificate. Provided only when the certificate type is AMAZON_ISSUED.

$sel:keyAlgorithm:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_keyAlgorithm - The algorithm that was used to generate the public-private key pair.

Valid values: RSA_2048 | RSA_1024 | RSA_4096 | EC_prime256v1 | EC_secp384r1 | EC_secp521r1

$sel:type':AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_type - The source of the certificate. For certificates that Certificate Manager provides, Type is AMAZON_ISSUED. For certificates that are imported with ImportCertificate, Type is IMPORTED.

Valid values: IMPORTED | AMAZON_ISSUED | PRIVATE

$sel:options:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_options - Provides a value that specifies whether to add the certificate to a transparency log.

$sel:issuedAt:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_issuedAt - Indicates when the certificate was issued. Provided if the certificate type is AMAZON_ISSUED.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:signatureAlgorithm:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_signatureAlgorithm - The algorithm that was used to sign the certificate.

$sel:domainValidationOptions:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_domainValidationOptions - Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request.

Only provided if the certificate type is AMAZON_ISSUED.

$sel:issuer:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_issuer - The name of the certificate authority that issued and signed the certificate.

$sel:notAfter:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_notAfter - The time after which the certificate becomes invalid.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:certificateAuthorityArn:AwsCertificateManagerCertificateDetails', awsCertificateManagerCertificateDetails_certificateAuthorityArn - The ARN of the private certificate authority (CA) that will be used to issue the certificate.

awsCertificateManagerCertificateDetails_subject :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

The name of the entity that is associated with the public key contained in the certificate.

awsCertificateManagerCertificateDetails_status :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

The status of the certificate.

Valid values: PENDING_VALIDATION | ISSUED | INACTIVE | EXPIRED | VALIDATION_TIMED_OUT | REVOKED | FAILED

awsCertificateManagerCertificateDetails_failureReason :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

For a failed certificate request, the reason for the failure.

Valid values: NO_AVAILABLE_CONTACTS | ADDITIONAL_VERIFICATION_REQUIRED | DOMAIN_NOT_ALLOWED | INVALID_PUBLIC_DOMAIN | DOMAIN_VALIDATION_DENIED | CAA_ERROR | PCA_LIMIT_EXCEEDED | PCA_INVALID_ARN | PCA_INVALID_STATE | PCA_REQUEST_FAILED | PCA_NAME_CONSTRAINTS_VALIDATION | PCA_RESOURCE_NOT_FOUND | PCA_INVALID_ARGS | PCA_INVALID_DURATION | PCA_ACCESS_DENIED | SLR_NOT_FOUND | OTHER

awsCertificateManagerCertificateDetails_subjectAlternativeNames :: Lens' AwsCertificateManagerCertificateDetails (Maybe [Text])

One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate.

The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

awsCertificateManagerCertificateDetails_inUseBy :: Lens' AwsCertificateManagerCertificateDetails (Maybe [Text])

The list of ARNs for the Amazon Web Services resources that use the certificate.

awsCertificateManagerCertificateDetails_createdAt :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

Indicates when the certificate was requested.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsCertificateManagerCertificateDetails_serial :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

The serial number of the certificate.

awsCertificateManagerCertificateDetails_renewalEligibility :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

Whether the certificate is eligible for renewal.

Valid values: ELIGIBLE | INELIGIBLE

awsCertificateManagerCertificateDetails_extendedKeyUsages :: Lens' AwsCertificateManagerCertificateDetails (Maybe [AwsCertificateManagerCertificateExtendedKeyUsage])

Contains a list of Extended Key Usage X.509 v3 extension objects. Each object specifies a purpose for which the certificate public key can be used and consists of a name and an object identifier (OID).

awsCertificateManagerCertificateDetails_importedAt :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

Indicates when the certificate was imported. Provided if the certificate type is IMPORTED.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsCertificateManagerCertificateDetails_keyUsages :: Lens' AwsCertificateManagerCertificateDetails (Maybe [AwsCertificateManagerCertificateKeyUsage])

A list of key usage X.509 v3 extension objects.

awsCertificateManagerCertificateDetails_notBefore :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

The time before which the certificate is not valid.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsCertificateManagerCertificateDetails_domainName :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

The fully qualified domain name (FQDN), such as www.example.com, that is secured by the certificate.

awsCertificateManagerCertificateDetails_renewalSummary :: Lens' AwsCertificateManagerCertificateDetails (Maybe AwsCertificateManagerCertificateRenewalSummary)

Information about the status of the Certificate Manager managed renewal for the certificate. Provided only when the certificate type is AMAZON_ISSUED.

awsCertificateManagerCertificateDetails_keyAlgorithm :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

The algorithm that was used to generate the public-private key pair.

Valid values: RSA_2048 | RSA_1024 | RSA_4096 | EC_prime256v1 | EC_secp384r1 | EC_secp521r1

awsCertificateManagerCertificateDetails_type :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

The source of the certificate. For certificates that Certificate Manager provides, Type is AMAZON_ISSUED. For certificates that are imported with ImportCertificate, Type is IMPORTED.

Valid values: IMPORTED | AMAZON_ISSUED | PRIVATE

awsCertificateManagerCertificateDetails_options :: Lens' AwsCertificateManagerCertificateDetails (Maybe AwsCertificateManagerCertificateOptions)

Provides a value that specifies whether to add the certificate to a transparency log.

awsCertificateManagerCertificateDetails_issuedAt :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

Indicates when the certificate was issued. Provided if the certificate type is AMAZON_ISSUED.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsCertificateManagerCertificateDetails_signatureAlgorithm :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

The algorithm that was used to sign the certificate.

awsCertificateManagerCertificateDetails_domainValidationOptions :: Lens' AwsCertificateManagerCertificateDetails (Maybe [AwsCertificateManagerCertificateDomainValidationOption])

Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request.

Only provided if the certificate type is AMAZON_ISSUED.

awsCertificateManagerCertificateDetails_issuer :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

The name of the certificate authority that issued and signed the certificate.

awsCertificateManagerCertificateDetails_notAfter :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

The time after which the certificate becomes invalid.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsCertificateManagerCertificateDetails_certificateAuthorityArn :: Lens' AwsCertificateManagerCertificateDetails (Maybe Text)

The ARN of the private certificate authority (CA) that will be used to issue the certificate.