Amazonka.SecurityHub.Types.AwsCertificateManagerCertificateExtendedKeyUsage

Description

Documentation

data AwsCertificateManagerCertificateExtendedKeyUsage

Contains information about an extended key usage X.509 v3 extension object.

See: newAwsCertificateManagerCertificateExtendedKeyUsage smart constructor.

Constructors

AwsCertificateManagerCertificateExtendedKeyUsage'

Fields

oId :: Maybe Text

An object identifier (OID) for the extension value.

The format is numbers separated by periods.

name :: Maybe Text

The name of an extension value. Indicates the purpose for which the certificate public key can be used.

newAwsCertificateManagerCertificateExtendedKeyUsage :: AwsCertificateManagerCertificateExtendedKeyUsage

Create a value of AwsCertificateManagerCertificateExtendedKeyUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:oId:AwsCertificateManagerCertificateExtendedKeyUsage', awsCertificateManagerCertificateExtendedKeyUsage_oId - An object identifier (OID) for the extension value.

The format is numbers separated by periods.

$sel:name:AwsCertificateManagerCertificateExtendedKeyUsage', awsCertificateManagerCertificateExtendedKeyUsage_name - The name of an extension value. Indicates the purpose for which the certificate public key can be used.

awsCertificateManagerCertificateExtendedKeyUsage_oId :: Lens' AwsCertificateManagerCertificateExtendedKeyUsage (Maybe Text)

An object identifier (OID) for the extension value.

The format is numbers separated by periods.

awsCertificateManagerCertificateExtendedKeyUsage_name :: Lens' AwsCertificateManagerCertificateExtendedKeyUsage (Maybe Text)

The name of an extension value. Indicates the purpose for which the certificate public key can be used.