Amazonka.SecurityHub.Types.AwsCertificateManagerCertificateRenewalSummary

Description

Documentation

data AwsCertificateManagerCertificateRenewalSummary

Contains information about the Certificate Manager managed renewal for an AMAZON_ISSUED certificate.

See: newAwsCertificateManagerCertificateRenewalSummary smart constructor.

Constructors

AwsCertificateManagerCertificateRenewalSummary'

Fields

renewalStatus :: Maybe Text

The status of the Certificate Manager managed renewal of the certificate.

Valid values: PENDING_AUTO_RENEWAL | PENDING_VALIDATION | SUCCESS | FAILED

updatedAt :: Maybe Text

Indicates when the renewal summary was last updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

domainValidationOptions :: Maybe [AwsCertificateManagerCertificateDomainValidationOption]

Information about the validation of each domain name in the certificate, as it pertains to Certificate Manager managed renewal. Provided only when the certificate type is AMAZON_ISSUED.

renewalStatusReason :: Maybe Text

The reason that a renewal request was unsuccessful.

Valid values: NO_AVAILABLE_CONTACTS | ADDITIONAL_VERIFICATION_REQUIRED | DOMAIN_NOT_ALLOWED | INVALID_PUBLIC_DOMAIN | DOMAIN_VALIDATION_DENIED | CAA_ERROR | PCA_LIMIT_EXCEEDED | PCA_INVALID_ARN | PCA_INVALID_STATE | PCA_REQUEST_FAILED | PCA_NAME_CONSTRAINTS_VALIDATION | PCA_RESOURCE_NOT_FOUND | PCA_INVALID_ARGS | PCA_INVALID_DURATION | PCA_ACCESS_DENIED | SLR_NOT_FOUND | OTHER

newAwsCertificateManagerCertificateRenewalSummary :: AwsCertificateManagerCertificateRenewalSummary

Create a value of AwsCertificateManagerCertificateRenewalSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renewalStatus:AwsCertificateManagerCertificateRenewalSummary', awsCertificateManagerCertificateRenewalSummary_renewalStatus - The status of the Certificate Manager managed renewal of the certificate.

Valid values: PENDING_AUTO_RENEWAL | PENDING_VALIDATION | SUCCESS | FAILED

$sel:updatedAt:AwsCertificateManagerCertificateRenewalSummary', awsCertificateManagerCertificateRenewalSummary_updatedAt - Indicates when the renewal summary was last updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:domainValidationOptions:AwsCertificateManagerCertificateRenewalSummary', awsCertificateManagerCertificateRenewalSummary_domainValidationOptions - Information about the validation of each domain name in the certificate, as it pertains to Certificate Manager managed renewal. Provided only when the certificate type is AMAZON_ISSUED.

$sel:renewalStatusReason:AwsCertificateManagerCertificateRenewalSummary', awsCertificateManagerCertificateRenewalSummary_renewalStatusReason - The reason that a renewal request was unsuccessful.

Valid values: NO_AVAILABLE_CONTACTS | ADDITIONAL_VERIFICATION_REQUIRED | DOMAIN_NOT_ALLOWED | INVALID_PUBLIC_DOMAIN | DOMAIN_VALIDATION_DENIED | CAA_ERROR | PCA_LIMIT_EXCEEDED | PCA_INVALID_ARN | PCA_INVALID_STATE | PCA_REQUEST_FAILED | PCA_NAME_CONSTRAINTS_VALIDATION | PCA_RESOURCE_NOT_FOUND | PCA_INVALID_ARGS | PCA_INVALID_DURATION | PCA_ACCESS_DENIED | SLR_NOT_FOUND | OTHER

awsCertificateManagerCertificateRenewalSummary_renewalStatus :: Lens' AwsCertificateManagerCertificateRenewalSummary (Maybe Text)

The status of the Certificate Manager managed renewal of the certificate.

Valid values: PENDING_AUTO_RENEWAL | PENDING_VALIDATION | SUCCESS | FAILED

awsCertificateManagerCertificateRenewalSummary_updatedAt :: Lens' AwsCertificateManagerCertificateRenewalSummary (Maybe Text)

Indicates when the renewal summary was last updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsCertificateManagerCertificateRenewalSummary_domainValidationOptions :: Lens' AwsCertificateManagerCertificateRenewalSummary (Maybe [AwsCertificateManagerCertificateDomainValidationOption])

Information about the validation of each domain name in the certificate, as it pertains to Certificate Manager managed renewal. Provided only when the certificate type is AMAZON_ISSUED.

awsCertificateManagerCertificateRenewalSummary_renewalStatusReason :: Lens' AwsCertificateManagerCertificateRenewalSummary (Maybe Text)

The reason that a renewal request was unsuccessful.

Valid values: NO_AVAILABLE_CONTACTS | ADDITIONAL_VERIFICATION_REQUIRED | DOMAIN_NOT_ALLOWED | INVALID_PUBLIC_DOMAIN | DOMAIN_VALIDATION_DENIED | CAA_ERROR | PCA_LIMIT_EXCEEDED | PCA_INVALID_ARN | PCA_INVALID_STATE | PCA_REQUEST_FAILED | PCA_NAME_CONSTRAINTS_VALIDATION | PCA_RESOURCE_NOT_FOUND | PCA_INVALID_ARGS | PCA_INVALID_DURATION | PCA_ACCESS_DENIED | SLR_NOT_FOUND | OTHER