Amazonka.SecurityHub.Types.AwsCloudFrontDistributionViewerCertificate

Description

Documentation

data AwsCloudFrontDistributionViewerCertificate

Provides information about the TLS/SSL configuration that the distribution uses to communicate with viewers.

See: newAwsCloudFrontDistributionViewerCertificate smart constructor.

Constructors

AwsCloudFrontDistributionViewerCertificate'

Fields

sslSupportMethod :: Maybe Text

The viewers that the distribution accepts HTTPS connections from.

acmCertificateArn :: Maybe Text

The ARN of the ACM certificate. Used if the certificate is stored in ACM. If you provide an ACM certificate ARN, you must also provide MinimumCertificateVersion and SslSupportMethod.

certificateSource :: Maybe Text

The source of the certificate identified by Certificate. Note that in CloudFront, this attribute is deprecated.

minimumProtocolVersion :: Maybe Text

The security policy that CloudFront uses for HTTPS connections with viewers. If SslSupportMethod is sni-only, then MinimumProtocolVersion must be TLSv1 or higher.

certificate :: Maybe Text

The identifier of the certificate. Note that in CloudFront, this attribute is deprecated.

iamCertificateId :: Maybe Text

The identifier of the IAM certificate. Used if the certificate is stored in IAM. If you provide IamCertificateId, then you also must provide MinimumProtocolVersion and SslSupportMethod.

cloudFrontDefaultCertificate :: Maybe Bool

Whether the distribution uses the CloudFront domain name. If set to false, then you provide either AcmCertificateArn or IamCertificateId.

newAwsCloudFrontDistributionViewerCertificate :: AwsCloudFrontDistributionViewerCertificate

Create a value of AwsCloudFrontDistributionViewerCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sslSupportMethod:AwsCloudFrontDistributionViewerCertificate', awsCloudFrontDistributionViewerCertificate_sslSupportMethod - The viewers that the distribution accepts HTTPS connections from.

$sel:acmCertificateArn:AwsCloudFrontDistributionViewerCertificate', awsCloudFrontDistributionViewerCertificate_acmCertificateArn - The ARN of the ACM certificate. Used if the certificate is stored in ACM. If you provide an ACM certificate ARN, you must also provide MinimumCertificateVersion and SslSupportMethod.

$sel:certificateSource:AwsCloudFrontDistributionViewerCertificate', awsCloudFrontDistributionViewerCertificate_certificateSource - The source of the certificate identified by Certificate. Note that in CloudFront, this attribute is deprecated.

$sel:minimumProtocolVersion:AwsCloudFrontDistributionViewerCertificate', awsCloudFrontDistributionViewerCertificate_minimumProtocolVersion - The security policy that CloudFront uses for HTTPS connections with viewers. If SslSupportMethod is sni-only, then MinimumProtocolVersion must be TLSv1 or higher.

$sel:certificate:AwsCloudFrontDistributionViewerCertificate', awsCloudFrontDistributionViewerCertificate_certificate - The identifier of the certificate. Note that in CloudFront, this attribute is deprecated.

$sel:iamCertificateId:AwsCloudFrontDistributionViewerCertificate', awsCloudFrontDistributionViewerCertificate_iamCertificateId - The identifier of the IAM certificate. Used if the certificate is stored in IAM. If you provide IamCertificateId, then you also must provide MinimumProtocolVersion and SslSupportMethod.

$sel:cloudFrontDefaultCertificate:AwsCloudFrontDistributionViewerCertificate', awsCloudFrontDistributionViewerCertificate_cloudFrontDefaultCertificate - Whether the distribution uses the CloudFront domain name. If set to false, then you provide either AcmCertificateArn or IamCertificateId.

awsCloudFrontDistributionViewerCertificate_sslSupportMethod :: Lens' AwsCloudFrontDistributionViewerCertificate (Maybe Text)

The viewers that the distribution accepts HTTPS connections from.

awsCloudFrontDistributionViewerCertificate_acmCertificateArn :: Lens' AwsCloudFrontDistributionViewerCertificate (Maybe Text)

The ARN of the ACM certificate. Used if the certificate is stored in ACM. If you provide an ACM certificate ARN, you must also provide MinimumCertificateVersion and SslSupportMethod.

awsCloudFrontDistributionViewerCertificate_certificateSource :: Lens' AwsCloudFrontDistributionViewerCertificate (Maybe Text)

The source of the certificate identified by Certificate. Note that in CloudFront, this attribute is deprecated.

awsCloudFrontDistributionViewerCertificate_minimumProtocolVersion :: Lens' AwsCloudFrontDistributionViewerCertificate (Maybe Text)

The security policy that CloudFront uses for HTTPS connections with viewers. If SslSupportMethod is sni-only, then MinimumProtocolVersion must be TLSv1 or higher.

awsCloudFrontDistributionViewerCertificate_certificate :: Lens' AwsCloudFrontDistributionViewerCertificate (Maybe Text)

The identifier of the certificate. Note that in CloudFront, this attribute is deprecated.

awsCloudFrontDistributionViewerCertificate_iamCertificateId :: Lens' AwsCloudFrontDistributionViewerCertificate (Maybe Text)

The identifier of the IAM certificate. Used if the certificate is stored in IAM. If you provide IamCertificateId, then you also must provide MinimumProtocolVersion and SslSupportMethod.

awsCloudFrontDistributionViewerCertificate_cloudFrontDefaultCertificate :: Lens' AwsCloudFrontDistributionViewerCertificate (Maybe Bool)

Whether the distribution uses the CloudFront domain name. If set to false, then you provide either AcmCertificateArn or IamCertificateId.