Amazonka.SecurityHub.Types.AwsCloudTrailTrailDetails

Description

Documentation

data AwsCloudTrailTrailDetails

Provides details about a CloudTrail trail.

See: newAwsCloudTrailTrailDetails smart constructor.

Constructors

AwsCloudTrailTrailDetails'

Fields

logFileValidationEnabled :: Maybe Bool

Indicates whether CloudTrail log file validation is enabled.

trailArn :: Maybe Text

The ARN of the trail.

s3KeyPrefix :: Maybe Text

The S3 key prefix. The key prefix is added after the name of the S3 bucket where the log files are published.

snsTopicArn :: Maybe Text

The ARN of the SNS topic that is used for notifications of log file delivery.

snsTopicName :: Maybe Text

The name of the SNS topic that is used for notifications of log file delivery.

cloudWatchLogsLogGroupArn :: Maybe Text

The ARN of the log group that CloudTrail logs are delivered to.

kmsKeyId :: Maybe Text

The KMS key ID to use to encrypt the logs.

homeRegion :: Maybe Text

The Region where the trail was created.

name :: Maybe Text

The name of the trail.

includeGlobalServiceEvents :: Maybe Bool

Indicates whether the trail publishes events from global services such as IAM to the log files.

hasCustomEventSelectors :: Maybe Bool

Indicates whether the trail has custom event selectors.

isOrganizationTrail :: Maybe Bool

Whether the trail is created for all accounts in an organization in Organizations, or only for the current Amazon Web Services account.

cloudWatchLogsRoleArn :: Maybe Text

The ARN of the role that the CloudWatch Events endpoint assumes when it writes to the log group.

s3BucketName :: Maybe Text

The name of the S3 bucket where the log files are published.

isMultiRegionTrail :: Maybe Bool

Indicates whether the trail applies only to the current Region or to all Regions.

newAwsCloudTrailTrailDetails :: AwsCloudTrailTrailDetails

Create a value of AwsCloudTrailTrailDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logFileValidationEnabled:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_logFileValidationEnabled - Indicates whether CloudTrail log file validation is enabled.

$sel:trailArn:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_trailArn - The ARN of the trail.

$sel:s3KeyPrefix:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_s3KeyPrefix - The S3 key prefix. The key prefix is added after the name of the S3 bucket where the log files are published.

$sel:snsTopicArn:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_snsTopicArn - The ARN of the SNS topic that is used for notifications of log file delivery.

$sel:snsTopicName:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_snsTopicName - The name of the SNS topic that is used for notifications of log file delivery.

$sel:cloudWatchLogsLogGroupArn:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_cloudWatchLogsLogGroupArn - The ARN of the log group that CloudTrail logs are delivered to.

$sel:kmsKeyId:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_kmsKeyId - The KMS key ID to use to encrypt the logs.

$sel:homeRegion:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_homeRegion - The Region where the trail was created.

$sel:name:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_name - The name of the trail.

$sel:includeGlobalServiceEvents:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_includeGlobalServiceEvents - Indicates whether the trail publishes events from global services such as IAM to the log files.

$sel:hasCustomEventSelectors:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_hasCustomEventSelectors - Indicates whether the trail has custom event selectors.

$sel:isOrganizationTrail:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_isOrganizationTrail - Whether the trail is created for all accounts in an organization in Organizations, or only for the current Amazon Web Services account.

$sel:cloudWatchLogsRoleArn:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_cloudWatchLogsRoleArn - The ARN of the role that the CloudWatch Events endpoint assumes when it writes to the log group.

$sel:s3BucketName:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_s3BucketName - The name of the S3 bucket where the log files are published.

$sel:isMultiRegionTrail:AwsCloudTrailTrailDetails', awsCloudTrailTrailDetails_isMultiRegionTrail - Indicates whether the trail applies only to the current Region or to all Regions.

awsCloudTrailTrailDetails_logFileValidationEnabled :: Lens' AwsCloudTrailTrailDetails (Maybe Bool)

Indicates whether CloudTrail log file validation is enabled.

awsCloudTrailTrailDetails_trailArn :: Lens' AwsCloudTrailTrailDetails (Maybe Text)

The ARN of the trail.

awsCloudTrailTrailDetails_s3KeyPrefix :: Lens' AwsCloudTrailTrailDetails (Maybe Text)

The S3 key prefix. The key prefix is added after the name of the S3 bucket where the log files are published.

awsCloudTrailTrailDetails_snsTopicArn :: Lens' AwsCloudTrailTrailDetails (Maybe Text)

The ARN of the SNS topic that is used for notifications of log file delivery.

awsCloudTrailTrailDetails_snsTopicName :: Lens' AwsCloudTrailTrailDetails (Maybe Text)

The name of the SNS topic that is used for notifications of log file delivery.

awsCloudTrailTrailDetails_cloudWatchLogsLogGroupArn :: Lens' AwsCloudTrailTrailDetails (Maybe Text)

The ARN of the log group that CloudTrail logs are delivered to.

awsCloudTrailTrailDetails_kmsKeyId :: Lens' AwsCloudTrailTrailDetails (Maybe Text)

The KMS key ID to use to encrypt the logs.

awsCloudTrailTrailDetails_homeRegion :: Lens' AwsCloudTrailTrailDetails (Maybe Text)

The Region where the trail was created.

awsCloudTrailTrailDetails_name :: Lens' AwsCloudTrailTrailDetails (Maybe Text)

The name of the trail.

awsCloudTrailTrailDetails_includeGlobalServiceEvents :: Lens' AwsCloudTrailTrailDetails (Maybe Bool)

Indicates whether the trail publishes events from global services such as IAM to the log files.

awsCloudTrailTrailDetails_hasCustomEventSelectors :: Lens' AwsCloudTrailTrailDetails (Maybe Bool)

Indicates whether the trail has custom event selectors.

awsCloudTrailTrailDetails_isOrganizationTrail :: Lens' AwsCloudTrailTrailDetails (Maybe Bool)

Whether the trail is created for all accounts in an organization in Organizations, or only for the current Amazon Web Services account.

awsCloudTrailTrailDetails_cloudWatchLogsRoleArn :: Lens' AwsCloudTrailTrailDetails (Maybe Text)

The ARN of the role that the CloudWatch Events endpoint assumes when it writes to the log group.

awsCloudTrailTrailDetails_s3BucketName :: Lens' AwsCloudTrailTrailDetails (Maybe Text)

The name of the S3 bucket where the log files are published.

awsCloudTrailTrailDetails_isMultiRegionTrail :: Lens' AwsCloudTrailTrailDetails (Maybe Bool)

Indicates whether the trail applies only to the current Region or to all Regions.