Amazonka.SecurityHub.Types.AwsCodeBuildProjectArtifactsDetails

Description

Documentation

data AwsCodeBuildProjectArtifactsDetails

Information about the build artifacts for the CodeBuild project.

See: newAwsCodeBuildProjectArtifactsDetails smart constructor.

Constructors

AwsCodeBuildProjectArtifactsDetails'

Fields

packaging :: Maybe Text

Only used when Type is S3. The type of output artifact to create.

location :: Maybe Text

Only used when Type is S3. The name of the S3 bucket where the artifact is located.

path :: Maybe Text

Only used when Type is S3. The path to the artifact. Used with Name and NamespaceType to determine the pattern for storing the artifact.

encryptionDisabled :: Maybe Bool

Indicates whether to disable encryption on the artifact. Only valid when Type is S3.

name :: Maybe Text

Only used when Type is S3. The name of the artifact. Used with NamepaceType and Path to determine the pattern for storing the artifact.

overrideArtifactName :: Maybe Bool

Whether the name specified in the buildspec file overrides the artifact name.

artifactIdentifier :: Maybe Text

An identifier for the artifact definition.

type' :: Maybe Text

The type of build artifact.

namespaceType :: Maybe Text

Only used when Type is S3. The value to use for the namespace. Used with Name and Path to determine the pattern for storing the artifact.

newAwsCodeBuildProjectArtifactsDetails :: AwsCodeBuildProjectArtifactsDetails

Create a value of AwsCodeBuildProjectArtifactsDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packaging:AwsCodeBuildProjectArtifactsDetails', awsCodeBuildProjectArtifactsDetails_packaging - Only used when Type is S3. The type of output artifact to create.

$sel:location:AwsCodeBuildProjectArtifactsDetails', awsCodeBuildProjectArtifactsDetails_location - Only used when Type is S3. The name of the S3 bucket where the artifact is located.

$sel:path:AwsCodeBuildProjectArtifactsDetails', awsCodeBuildProjectArtifactsDetails_path - Only used when Type is S3. The path to the artifact. Used with Name and NamespaceType to determine the pattern for storing the artifact.

$sel:encryptionDisabled:AwsCodeBuildProjectArtifactsDetails', awsCodeBuildProjectArtifactsDetails_encryptionDisabled - Indicates whether to disable encryption on the artifact. Only valid when Type is S3.

$sel:name:AwsCodeBuildProjectArtifactsDetails', awsCodeBuildProjectArtifactsDetails_name - Only used when Type is S3. The name of the artifact. Used with NamepaceType and Path to determine the pattern for storing the artifact.

$sel:overrideArtifactName:AwsCodeBuildProjectArtifactsDetails', awsCodeBuildProjectArtifactsDetails_overrideArtifactName - Whether the name specified in the buildspec file overrides the artifact name.

$sel:artifactIdentifier:AwsCodeBuildProjectArtifactsDetails', awsCodeBuildProjectArtifactsDetails_artifactIdentifier - An identifier for the artifact definition.

$sel:type':AwsCodeBuildProjectArtifactsDetails', awsCodeBuildProjectArtifactsDetails_type - The type of build artifact.

$sel:namespaceType:AwsCodeBuildProjectArtifactsDetails', awsCodeBuildProjectArtifactsDetails_namespaceType - Only used when Type is S3. The value to use for the namespace. Used with Name and Path to determine the pattern for storing the artifact.

awsCodeBuildProjectArtifactsDetails_packaging :: Lens' AwsCodeBuildProjectArtifactsDetails (Maybe Text)

Only used when Type is S3. The type of output artifact to create.

awsCodeBuildProjectArtifactsDetails_location :: Lens' AwsCodeBuildProjectArtifactsDetails (Maybe Text)

Only used when Type is S3. The name of the S3 bucket where the artifact is located.

awsCodeBuildProjectArtifactsDetails_path :: Lens' AwsCodeBuildProjectArtifactsDetails (Maybe Text)

Only used when Type is S3. The path to the artifact. Used with Name and NamespaceType to determine the pattern for storing the artifact.

awsCodeBuildProjectArtifactsDetails_encryptionDisabled :: Lens' AwsCodeBuildProjectArtifactsDetails (Maybe Bool)

Indicates whether to disable encryption on the artifact. Only valid when Type is S3.

awsCodeBuildProjectArtifactsDetails_name :: Lens' AwsCodeBuildProjectArtifactsDetails (Maybe Text)

Only used when Type is S3. The name of the artifact. Used with NamepaceType and Path to determine the pattern for storing the artifact.

awsCodeBuildProjectArtifactsDetails_overrideArtifactName :: Lens' AwsCodeBuildProjectArtifactsDetails (Maybe Bool)

Whether the name specified in the buildspec file overrides the artifact name.

awsCodeBuildProjectArtifactsDetails_artifactIdentifier :: Lens' AwsCodeBuildProjectArtifactsDetails (Maybe Text)

An identifier for the artifact definition.

awsCodeBuildProjectArtifactsDetails_type :: Lens' AwsCodeBuildProjectArtifactsDetails (Maybe Text)

The type of build artifact.

awsCodeBuildProjectArtifactsDetails_namespaceType :: Lens' AwsCodeBuildProjectArtifactsDetails (Maybe Text)

Only used when Type is S3. The value to use for the namespace. Used with Name and Path to determine the pattern for storing the artifact.