Amazonka.SecurityHub.Types.AwsCodeBuildProjectVpcConfig

Description

Documentation

data AwsCodeBuildProjectVpcConfig

Information about the VPC configuration that CodeBuild accesses.

See: newAwsCodeBuildProjectVpcConfig smart constructor.

Constructors

AwsCodeBuildProjectVpcConfig'

Fields

securityGroupIds :: Maybe [Text]

A list of one or more security group IDs in your VPC.

vpcId :: Maybe Text

The ID of the VPC.

subnets :: Maybe [Text]

A list of one or more subnet IDs in your VPC.

newAwsCodeBuildProjectVpcConfig :: AwsCodeBuildProjectVpcConfig

Create a value of AwsCodeBuildProjectVpcConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:AwsCodeBuildProjectVpcConfig', awsCodeBuildProjectVpcConfig_securityGroupIds - A list of one or more security group IDs in your VPC.

$sel:vpcId:AwsCodeBuildProjectVpcConfig', awsCodeBuildProjectVpcConfig_vpcId - The ID of the VPC.

$sel:subnets:AwsCodeBuildProjectVpcConfig', awsCodeBuildProjectVpcConfig_subnets - A list of one or more subnet IDs in your VPC.

awsCodeBuildProjectVpcConfig_securityGroupIds :: Lens' AwsCodeBuildProjectVpcConfig (Maybe [Text])

A list of one or more security group IDs in your VPC.

awsCodeBuildProjectVpcConfig_vpcId :: Lens' AwsCodeBuildProjectVpcConfig (Maybe Text)

The ID of the VPC.

awsCodeBuildProjectVpcConfig_subnets :: Lens' AwsCodeBuildProjectVpcConfig (Maybe [Text])

A list of one or more subnet IDs in your VPC.