Amazonka.SecurityHub.Types.AwsCorsConfiguration

Description

Documentation

data AwsCorsConfiguration

Contains the cross-origin resource sharing (CORS) configuration for the API. CORS is only supported for HTTP APIs.

See: newAwsCorsConfiguration smart constructor.

Constructors

AwsCorsConfiguration'

Fields

maxAge :: Maybe Int

The number of seconds for which the browser caches preflight request results.

allowMethods :: Maybe [Text]

The allowed methods for CORS requests.

allowHeaders :: Maybe [Text]

The allowed headers for CORS requests.

exposeHeaders :: Maybe [Text]

The exposed headers for CORS requests.

allowOrigins :: Maybe [Text]

The allowed origins for CORS requests.

allowCredentials :: Maybe Bool

Indicates whether the CORS request includes credentials.

newAwsCorsConfiguration :: AwsCorsConfiguration

Create a value of AwsCorsConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxAge:AwsCorsConfiguration', awsCorsConfiguration_maxAge - The number of seconds for which the browser caches preflight request results.

$sel:allowMethods:AwsCorsConfiguration', awsCorsConfiguration_allowMethods - The allowed methods for CORS requests.

$sel:allowHeaders:AwsCorsConfiguration', awsCorsConfiguration_allowHeaders - The allowed headers for CORS requests.

$sel:exposeHeaders:AwsCorsConfiguration', awsCorsConfiguration_exposeHeaders - The exposed headers for CORS requests.

$sel:allowOrigins:AwsCorsConfiguration', awsCorsConfiguration_allowOrigins - The allowed origins for CORS requests.

$sel:allowCredentials:AwsCorsConfiguration', awsCorsConfiguration_allowCredentials - Indicates whether the CORS request includes credentials.

awsCorsConfiguration_maxAge :: Lens' AwsCorsConfiguration (Maybe Int)

The number of seconds for which the browser caches preflight request results.

awsCorsConfiguration_allowMethods :: Lens' AwsCorsConfiguration (Maybe [Text])

The allowed methods for CORS requests.

awsCorsConfiguration_allowHeaders :: Lens' AwsCorsConfiguration (Maybe [Text])

The allowed headers for CORS requests.

awsCorsConfiguration_exposeHeaders :: Lens' AwsCorsConfiguration (Maybe [Text])

The exposed headers for CORS requests.

awsCorsConfiguration_allowOrigins :: Lens' AwsCorsConfiguration (Maybe [Text])

The allowed origins for CORS requests.

awsCorsConfiguration_allowCredentials :: Lens' AwsCorsConfiguration (Maybe Bool)

Indicates whether the CORS request includes credentials.