Amazonka.SecurityHub.Types.AwsDynamoDbTableDetails

Description

Documentation

data AwsDynamoDbTableDetails

Provides details about a DynamoDB table.

See: newAwsDynamoDbTableDetails smart constructor.

Constructors

AwsDynamoDbTableDetails'

Fields

restoreSummary :: Maybe AwsDynamoDbTableRestoreSummary

Information about the restore for the table.

globalTableVersion :: Maybe Text

The version of global tables being used.

tableSizeBytes :: Maybe Integer

The total size of the table in bytes.

attributeDefinitions :: Maybe [AwsDynamoDbTableAttributeDefinition]

A list of attribute definitions for the table.

latestStreamArn :: Maybe Text

The ARN of the latest stream for the table.

provisionedThroughput :: Maybe AwsDynamoDbTableProvisionedThroughput

Information about the provisioned throughput for the table.

tableStatus :: Maybe Text

The current status of the table.

keySchema :: Maybe [AwsDynamoDbTableKeySchema]

The primary key structure for the table.

globalSecondaryIndexes :: Maybe [AwsDynamoDbTableGlobalSecondaryIndex]

List of global secondary indexes for the table.

latestStreamLabel :: Maybe Text

The label of the latest stream. The label is not a unique identifier.

billingModeSummary :: Maybe AwsDynamoDbTableBillingModeSummary

Information about the billing for read/write capacity on the table.

localSecondaryIndexes :: Maybe [AwsDynamoDbTableLocalSecondaryIndex]

The list of local secondary indexes for the table.

creationDateTime :: Maybe Text

Indicates when the table was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

sseDescription :: Maybe AwsDynamoDbTableSseDescription

Information about the server-side encryption for the table.

tableId :: Maybe Text

The identifier of the table.

replicas :: Maybe [AwsDynamoDbTableReplica]

The list of replicas of this table.

itemCount :: Maybe Int

The number of items in the table.

tableName :: Maybe Text

The name of the table.

streamSpecification :: Maybe AwsDynamoDbTableStreamSpecification

The current DynamoDB Streams configuration for the table.

newAwsDynamoDbTableDetails :: AwsDynamoDbTableDetails

Create a value of AwsDynamoDbTableDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:restoreSummary:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_restoreSummary - Information about the restore for the table.

$sel:globalTableVersion:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_globalTableVersion - The version of global tables being used.

$sel:tableSizeBytes:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_tableSizeBytes - The total size of the table in bytes.

$sel:attributeDefinitions:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_attributeDefinitions - A list of attribute definitions for the table.

$sel:latestStreamArn:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_latestStreamArn - The ARN of the latest stream for the table.

$sel:provisionedThroughput:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_provisionedThroughput - Information about the provisioned throughput for the table.

$sel:tableStatus:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_tableStatus - The current status of the table.

$sel:keySchema:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_keySchema - The primary key structure for the table.

$sel:globalSecondaryIndexes:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_globalSecondaryIndexes - List of global secondary indexes for the table.

$sel:latestStreamLabel:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_latestStreamLabel - The label of the latest stream. The label is not a unique identifier.

$sel:billingModeSummary:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_billingModeSummary - Information about the billing for read/write capacity on the table.

$sel:localSecondaryIndexes:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_localSecondaryIndexes - The list of local secondary indexes for the table.

$sel:creationDateTime:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_creationDateTime - Indicates when the table was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:sseDescription:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_sseDescription - Information about the server-side encryption for the table.

$sel:tableId:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_tableId - The identifier of the table.

$sel:replicas:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_replicas - The list of replicas of this table.

$sel:itemCount:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_itemCount - The number of items in the table.

$sel:tableName:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_tableName - The name of the table.

$sel:streamSpecification:AwsDynamoDbTableDetails', awsDynamoDbTableDetails_streamSpecification - The current DynamoDB Streams configuration for the table.

awsDynamoDbTableDetails_restoreSummary :: Lens' AwsDynamoDbTableDetails (Maybe AwsDynamoDbTableRestoreSummary)

Information about the restore for the table.

awsDynamoDbTableDetails_globalTableVersion :: Lens' AwsDynamoDbTableDetails (Maybe Text)

The version of global tables being used.

awsDynamoDbTableDetails_tableSizeBytes :: Lens' AwsDynamoDbTableDetails (Maybe Integer)

The total size of the table in bytes.

awsDynamoDbTableDetails_attributeDefinitions :: Lens' AwsDynamoDbTableDetails (Maybe [AwsDynamoDbTableAttributeDefinition])

A list of attribute definitions for the table.

awsDynamoDbTableDetails_latestStreamArn :: Lens' AwsDynamoDbTableDetails (Maybe Text)

The ARN of the latest stream for the table.

awsDynamoDbTableDetails_provisionedThroughput :: Lens' AwsDynamoDbTableDetails (Maybe AwsDynamoDbTableProvisionedThroughput)

Information about the provisioned throughput for the table.

awsDynamoDbTableDetails_tableStatus :: Lens' AwsDynamoDbTableDetails (Maybe Text)

The current status of the table.

awsDynamoDbTableDetails_keySchema :: Lens' AwsDynamoDbTableDetails (Maybe [AwsDynamoDbTableKeySchema])

The primary key structure for the table.

awsDynamoDbTableDetails_globalSecondaryIndexes :: Lens' AwsDynamoDbTableDetails (Maybe [AwsDynamoDbTableGlobalSecondaryIndex])

List of global secondary indexes for the table.

awsDynamoDbTableDetails_latestStreamLabel :: Lens' AwsDynamoDbTableDetails (Maybe Text)

The label of the latest stream. The label is not a unique identifier.

awsDynamoDbTableDetails_billingModeSummary :: Lens' AwsDynamoDbTableDetails (Maybe AwsDynamoDbTableBillingModeSummary)

Information about the billing for read/write capacity on the table.

awsDynamoDbTableDetails_localSecondaryIndexes :: Lens' AwsDynamoDbTableDetails (Maybe [AwsDynamoDbTableLocalSecondaryIndex])

The list of local secondary indexes for the table.

awsDynamoDbTableDetails_creationDateTime :: Lens' AwsDynamoDbTableDetails (Maybe Text)

Indicates when the table was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsDynamoDbTableDetails_sseDescription :: Lens' AwsDynamoDbTableDetails (Maybe AwsDynamoDbTableSseDescription)

Information about the server-side encryption for the table.

awsDynamoDbTableDetails_tableId :: Lens' AwsDynamoDbTableDetails (Maybe Text)

The identifier of the table.

awsDynamoDbTableDetails_replicas :: Lens' AwsDynamoDbTableDetails (Maybe [AwsDynamoDbTableReplica])

The list of replicas of this table.

awsDynamoDbTableDetails_itemCount :: Lens' AwsDynamoDbTableDetails (Maybe Int)

The number of items in the table.

awsDynamoDbTableDetails_tableName :: Lens' AwsDynamoDbTableDetails (Maybe Text)

The name of the table.

awsDynamoDbTableDetails_streamSpecification :: Lens' AwsDynamoDbTableDetails (Maybe AwsDynamoDbTableStreamSpecification)

The current DynamoDB Streams configuration for the table.