Amazonka.SecurityHub.Types.AwsDynamoDbTableGlobalSecondaryIndex

Description

Documentation

data AwsDynamoDbTableGlobalSecondaryIndex

Information abut a global secondary index for the table.

See: newAwsDynamoDbTableGlobalSecondaryIndex smart constructor.

Constructors

AwsDynamoDbTableGlobalSecondaryIndex'

Fields

backfilling :: Maybe Bool

Whether the index is currently backfilling.

indexSizeBytes :: Maybe Integer

The total size in bytes of the index.

indexStatus :: Maybe Text

The current status of the index.

provisionedThroughput :: Maybe AwsDynamoDbTableProvisionedThroughput

Information about the provisioned throughput settings for the indexes.

indexArn :: Maybe Text

The ARN of the index.

keySchema :: Maybe [AwsDynamoDbTableKeySchema]

The key schema for the index.

projection :: Maybe AwsDynamoDbTableProjection

Attributes that are copied from the table into an index.

itemCount :: Maybe Int

The number of items in the index.

indexName :: Maybe Text

The name of the index.

newAwsDynamoDbTableGlobalSecondaryIndex :: AwsDynamoDbTableGlobalSecondaryIndex

Create a value of AwsDynamoDbTableGlobalSecondaryIndex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backfilling:AwsDynamoDbTableGlobalSecondaryIndex', awsDynamoDbTableGlobalSecondaryIndex_backfilling - Whether the index is currently backfilling.

$sel:indexSizeBytes:AwsDynamoDbTableGlobalSecondaryIndex', awsDynamoDbTableGlobalSecondaryIndex_indexSizeBytes - The total size in bytes of the index.

$sel:indexStatus:AwsDynamoDbTableGlobalSecondaryIndex', awsDynamoDbTableGlobalSecondaryIndex_indexStatus - The current status of the index.

$sel:provisionedThroughput:AwsDynamoDbTableGlobalSecondaryIndex', awsDynamoDbTableGlobalSecondaryIndex_provisionedThroughput - Information about the provisioned throughput settings for the indexes.

$sel:indexArn:AwsDynamoDbTableGlobalSecondaryIndex', awsDynamoDbTableGlobalSecondaryIndex_indexArn - The ARN of the index.

$sel:keySchema:AwsDynamoDbTableGlobalSecondaryIndex', awsDynamoDbTableGlobalSecondaryIndex_keySchema - The key schema for the index.

$sel:projection:AwsDynamoDbTableGlobalSecondaryIndex', awsDynamoDbTableGlobalSecondaryIndex_projection - Attributes that are copied from the table into an index.

$sel:itemCount:AwsDynamoDbTableGlobalSecondaryIndex', awsDynamoDbTableGlobalSecondaryIndex_itemCount - The number of items in the index.

$sel:indexName:AwsDynamoDbTableGlobalSecondaryIndex', awsDynamoDbTableGlobalSecondaryIndex_indexName - The name of the index.

awsDynamoDbTableGlobalSecondaryIndex_backfilling :: Lens' AwsDynamoDbTableGlobalSecondaryIndex (Maybe Bool)

Whether the index is currently backfilling.

awsDynamoDbTableGlobalSecondaryIndex_indexSizeBytes :: Lens' AwsDynamoDbTableGlobalSecondaryIndex (Maybe Integer)

The total size in bytes of the index.

awsDynamoDbTableGlobalSecondaryIndex_indexStatus :: Lens' AwsDynamoDbTableGlobalSecondaryIndex (Maybe Text)

The current status of the index.

awsDynamoDbTableGlobalSecondaryIndex_provisionedThroughput :: Lens' AwsDynamoDbTableGlobalSecondaryIndex (Maybe AwsDynamoDbTableProvisionedThroughput)

Information about the provisioned throughput settings for the indexes.

awsDynamoDbTableGlobalSecondaryIndex_indexArn :: Lens' AwsDynamoDbTableGlobalSecondaryIndex (Maybe Text)

The ARN of the index.

awsDynamoDbTableGlobalSecondaryIndex_keySchema :: Lens' AwsDynamoDbTableGlobalSecondaryIndex (Maybe [AwsDynamoDbTableKeySchema])

The key schema for the index.

awsDynamoDbTableGlobalSecondaryIndex_projection :: Lens' AwsDynamoDbTableGlobalSecondaryIndex (Maybe AwsDynamoDbTableProjection)

Attributes that are copied from the table into an index.

awsDynamoDbTableGlobalSecondaryIndex_itemCount :: Lens' AwsDynamoDbTableGlobalSecondaryIndex (Maybe Int)

The number of items in the index.

awsDynamoDbTableGlobalSecondaryIndex_indexName :: Lens' AwsDynamoDbTableGlobalSecondaryIndex (Maybe Text)

The name of the index.