Amazonka.SecurityHub.Types.AwsDynamoDbTableLocalSecondaryIndex

Description

Documentation

data AwsDynamoDbTableLocalSecondaryIndex

Information about a local secondary index for a DynamoDB table.

See: newAwsDynamoDbTableLocalSecondaryIndex smart constructor.

Constructors

AwsDynamoDbTableLocalSecondaryIndex'

Fields

indexArn :: Maybe Text

The ARN of the index.

keySchema :: Maybe [AwsDynamoDbTableKeySchema]

The complete key schema for the index.

projection :: Maybe AwsDynamoDbTableProjection

Attributes that are copied from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

indexName :: Maybe Text

The name of the index.

newAwsDynamoDbTableLocalSecondaryIndex :: AwsDynamoDbTableLocalSecondaryIndex

Create a value of AwsDynamoDbTableLocalSecondaryIndex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:indexArn:AwsDynamoDbTableLocalSecondaryIndex', awsDynamoDbTableLocalSecondaryIndex_indexArn - The ARN of the index.

$sel:keySchema:AwsDynamoDbTableLocalSecondaryIndex', awsDynamoDbTableLocalSecondaryIndex_keySchema - The complete key schema for the index.

$sel:projection:AwsDynamoDbTableLocalSecondaryIndex', awsDynamoDbTableLocalSecondaryIndex_projection - Attributes that are copied from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

$sel:indexName:AwsDynamoDbTableLocalSecondaryIndex', awsDynamoDbTableLocalSecondaryIndex_indexName - The name of the index.

awsDynamoDbTableLocalSecondaryIndex_indexArn :: Lens' AwsDynamoDbTableLocalSecondaryIndex (Maybe Text)

The ARN of the index.

awsDynamoDbTableLocalSecondaryIndex_keySchema :: Lens' AwsDynamoDbTableLocalSecondaryIndex (Maybe [AwsDynamoDbTableKeySchema])

The complete key schema for the index.

awsDynamoDbTableLocalSecondaryIndex_projection :: Lens' AwsDynamoDbTableLocalSecondaryIndex (Maybe AwsDynamoDbTableProjection)

Attributes that are copied from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

awsDynamoDbTableLocalSecondaryIndex_indexName :: Lens' AwsDynamoDbTableLocalSecondaryIndex (Maybe Text)

The name of the index.