Amazonka.SecurityHub.Types.AwsEc2InstanceDetails

Description

Documentation

data AwsEc2InstanceDetails

The details of an EC2 instance.

See: newAwsEc2InstanceDetails smart constructor.

Constructors

AwsEc2InstanceDetails'

Fields

iamInstanceProfileArn :: Maybe Text

The IAM profile ARN of the instance.

vpcId :: Maybe Text

The identifier of the VPC that the instance was launched in.

keyName :: Maybe Text

The key name associated with the instance.

networkInterfaces :: Maybe [AwsEc2InstanceNetworkInterfacesDetails]

The identifiers of the network interfaces for the EC2 instance. The details for each network interface are in a corresponding AwsEc2NetworkInterfacesDetails object.

ipV4Addresses :: Maybe [Text]

The IPv4 addresses associated with the instance.

subnetId :: Maybe Text

The identifier of the subnet that the instance was launched in.

imageId :: Maybe Text

The Amazon Machine Image (AMI) ID of the instance.

type' :: Maybe Text

The instance type of the instance.

ipV6Addresses :: Maybe [Text]

The IPv6 addresses associated with the instance.

launchedAt :: Maybe Text

Indicates when the instance was launched.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

newAwsEc2InstanceDetails :: AwsEc2InstanceDetails

Create a value of AwsEc2InstanceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamInstanceProfileArn:AwsEc2InstanceDetails', awsEc2InstanceDetails_iamInstanceProfileArn - The IAM profile ARN of the instance.

$sel:vpcId:AwsEc2InstanceDetails', awsEc2InstanceDetails_vpcId - The identifier of the VPC that the instance was launched in.

$sel:keyName:AwsEc2InstanceDetails', awsEc2InstanceDetails_keyName - The key name associated with the instance.

$sel:networkInterfaces:AwsEc2InstanceDetails', awsEc2InstanceDetails_networkInterfaces - The identifiers of the network interfaces for the EC2 instance. The details for each network interface are in a corresponding AwsEc2NetworkInterfacesDetails object.

$sel:ipV4Addresses:AwsEc2InstanceDetails', awsEc2InstanceDetails_ipV4Addresses - The IPv4 addresses associated with the instance.

$sel:subnetId:AwsEc2InstanceDetails', awsEc2InstanceDetails_subnetId - The identifier of the subnet that the instance was launched in.

$sel:imageId:AwsEc2InstanceDetails', awsEc2InstanceDetails_imageId - The Amazon Machine Image (AMI) ID of the instance.

$sel:type':AwsEc2InstanceDetails', awsEc2InstanceDetails_type - The instance type of the instance.

$sel:ipV6Addresses:AwsEc2InstanceDetails', awsEc2InstanceDetails_ipV6Addresses - The IPv6 addresses associated with the instance.

$sel:launchedAt:AwsEc2InstanceDetails', awsEc2InstanceDetails_launchedAt - Indicates when the instance was launched.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsEc2InstanceDetails_iamInstanceProfileArn :: Lens' AwsEc2InstanceDetails (Maybe Text)

The IAM profile ARN of the instance.

awsEc2InstanceDetails_vpcId :: Lens' AwsEc2InstanceDetails (Maybe Text)

The identifier of the VPC that the instance was launched in.

awsEc2InstanceDetails_keyName :: Lens' AwsEc2InstanceDetails (Maybe Text)

The key name associated with the instance.

awsEc2InstanceDetails_networkInterfaces :: Lens' AwsEc2InstanceDetails (Maybe [AwsEc2InstanceNetworkInterfacesDetails])

The identifiers of the network interfaces for the EC2 instance. The details for each network interface are in a corresponding AwsEc2NetworkInterfacesDetails object.

awsEc2InstanceDetails_ipV4Addresses :: Lens' AwsEc2InstanceDetails (Maybe [Text])

The IPv4 addresses associated with the instance.

awsEc2InstanceDetails_subnetId :: Lens' AwsEc2InstanceDetails (Maybe Text)

The identifier of the subnet that the instance was launched in.

awsEc2InstanceDetails_imageId :: Lens' AwsEc2InstanceDetails (Maybe Text)

The Amazon Machine Image (AMI) ID of the instance.

awsEc2InstanceDetails_type :: Lens' AwsEc2InstanceDetails (Maybe Text)

The instance type of the instance.

awsEc2InstanceDetails_ipV6Addresses :: Lens' AwsEc2InstanceDetails (Maybe [Text])

The IPv6 addresses associated with the instance.

awsEc2InstanceDetails_launchedAt :: Lens' AwsEc2InstanceDetails (Maybe Text)

Indicates when the instance was launched.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.