Amazonka.SecurityHub.Types.AwsEc2SecurityGroupIpPermission

Description

Documentation

data AwsEc2SecurityGroupIpPermission

An IP permission for an EC2 security group.

See: newAwsEc2SecurityGroupIpPermission smart constructor.

Constructors

AwsEc2SecurityGroupIpPermission'

Fields

fromPort :: Maybe Int

The start of the port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type number.

A value of -1 indicates all ICMP/ICMPv6 types. If you specify all ICMP/ICMPv6 types, you must specify all codes.

userIdGroupPairs :: Maybe [AwsEc2SecurityGroupUserIdGroupPair]

The security group and Amazon Web Services account ID pairs.

prefixListIds :: Maybe [AwsEc2SecurityGroupPrefixListId]

VPC only

The prefix list IDs for an Amazon Web Services service. With outbound rules, this is the Amazon Web Services service to access through a VPC endpoint from instances associated with the security group.

ipProtocol :: Maybe Text

The IP protocol name (tcp, udp, icmp, icmpv6) or number.

VPC only

Use -1 to specify all protocols.

When authorizing security group rules, specifying -1 or a protocol number other than tcp, udp, icmp, or icmpv6 allows traffic on all ports, regardless of any port range you specify.

For tcp, udp, and icmp, you must specify a port range.

For icmpv6, the port range is optional. If you omit the port range, traffic for all types and codes is allowed.

toPort :: Maybe Int

The end of the port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code.

A value of -1 indicates all ICMP/ICMPv6 codes. If you specify all ICMP/ICMPv6 types, you must specify all codes.

ipv6Ranges :: Maybe [AwsEc2SecurityGroupIpv6Range]

The IPv6 ranges.

ipRanges :: Maybe [AwsEc2SecurityGroupIpRange]

The IPv4 ranges.

newAwsEc2SecurityGroupIpPermission :: AwsEc2SecurityGroupIpPermission

Create a value of AwsEc2SecurityGroupIpPermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromPort:AwsEc2SecurityGroupIpPermission', awsEc2SecurityGroupIpPermission_fromPort - The start of the port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type number.

A value of -1 indicates all ICMP/ICMPv6 types. If you specify all ICMP/ICMPv6 types, you must specify all codes.

$sel:userIdGroupPairs:AwsEc2SecurityGroupIpPermission', awsEc2SecurityGroupIpPermission_userIdGroupPairs - The security group and Amazon Web Services account ID pairs.

$sel:prefixListIds:AwsEc2SecurityGroupIpPermission', awsEc2SecurityGroupIpPermission_prefixListIds - [VPC only] The prefix list IDs for an Amazon Web Services service. With outbound rules, this is the Amazon Web Services service to access through a VPC endpoint from instances associated with the security group.

$sel:ipProtocol:AwsEc2SecurityGroupIpPermission', awsEc2SecurityGroupIpPermission_ipProtocol - The IP protocol name (tcp, udp, icmp, icmpv6) or number.

VPC only

Use -1 to specify all protocols.

When authorizing security group rules, specifying -1 or a protocol number other than tcp, udp, icmp, or icmpv6 allows traffic on all ports, regardless of any port range you specify.

For tcp, udp, and icmp, you must specify a port range.

For icmpv6, the port range is optional. If you omit the port range, traffic for all types and codes is allowed.

$sel:toPort:AwsEc2SecurityGroupIpPermission', awsEc2SecurityGroupIpPermission_toPort - The end of the port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code.

A value of -1 indicates all ICMP/ICMPv6 codes. If you specify all ICMP/ICMPv6 types, you must specify all codes.

$sel:ipv6Ranges:AwsEc2SecurityGroupIpPermission', awsEc2SecurityGroupIpPermission_ipv6Ranges - The IPv6 ranges.

$sel:ipRanges:AwsEc2SecurityGroupIpPermission', awsEc2SecurityGroupIpPermission_ipRanges - The IPv4 ranges.

awsEc2SecurityGroupIpPermission_fromPort :: Lens' AwsEc2SecurityGroupIpPermission (Maybe Int)

The start of the port range for the TCP and UDP protocols, or an ICMP/ICMPv6 type number.

A value of -1 indicates all ICMP/ICMPv6 types. If you specify all ICMP/ICMPv6 types, you must specify all codes.

awsEc2SecurityGroupIpPermission_userIdGroupPairs :: Lens' AwsEc2SecurityGroupIpPermission (Maybe [AwsEc2SecurityGroupUserIdGroupPair])

The security group and Amazon Web Services account ID pairs.

awsEc2SecurityGroupIpPermission_prefixListIds :: Lens' AwsEc2SecurityGroupIpPermission (Maybe [AwsEc2SecurityGroupPrefixListId])

VPC only

The prefix list IDs for an Amazon Web Services service. With outbound rules, this is the Amazon Web Services service to access through a VPC endpoint from instances associated with the security group.

awsEc2SecurityGroupIpPermission_ipProtocol :: Lens' AwsEc2SecurityGroupIpPermission (Maybe Text)

The IP protocol name (tcp, udp, icmp, icmpv6) or number.

VPC only

Use -1 to specify all protocols.

When authorizing security group rules, specifying -1 or a protocol number other than tcp, udp, icmp, or icmpv6 allows traffic on all ports, regardless of any port range you specify.

For tcp, udp, and icmp, you must specify a port range.

For icmpv6, the port range is optional. If you omit the port range, traffic for all types and codes is allowed.

awsEc2SecurityGroupIpPermission_toPort :: Lens' AwsEc2SecurityGroupIpPermission (Maybe Int)

The end of the port range for the TCP and UDP protocols, or an ICMP/ICMPv6 code.

A value of -1 indicates all ICMP/ICMPv6 codes. If you specify all ICMP/ICMPv6 types, you must specify all codes.

awsEc2SecurityGroupIpPermission_ipv6Ranges :: Lens' AwsEc2SecurityGroupIpPermission (Maybe [AwsEc2SecurityGroupIpv6Range])

The IPv6 ranges.

awsEc2SecurityGroupIpPermission_ipRanges :: Lens' AwsEc2SecurityGroupIpPermission (Maybe [AwsEc2SecurityGroupIpRange])

The IPv4 ranges.