Amazonka.SecurityHub.Types.AwsEc2SecurityGroupUserIdGroupPair

Description

Documentation

data AwsEc2SecurityGroupUserIdGroupPair

A relationship between a security group and a user.

See: newAwsEc2SecurityGroupUserIdGroupPair smart constructor.

Constructors

AwsEc2SecurityGroupUserIdGroupPair'

Fields

vpcPeeringConnectionId :: Maybe Text

The ID of the VPC peering connection, if applicable.

vpcId :: Maybe Text

The ID of the VPC for the referenced security group, if applicable.

userId :: Maybe Text

The ID of an Amazon Web Services account.

For a referenced security group in another VPC, the account ID of the referenced security group is returned in the response. If the referenced security group is deleted, this value is not returned.

EC2-Classic

Required when adding or removing rules that reference a security group in another VPC.

groupId :: Maybe Text

The ID of the security group.

groupName :: Maybe Text

The name of the security group.

peeringStatus :: Maybe Text

The status of a VPC peering connection, if applicable.

newAwsEc2SecurityGroupUserIdGroupPair :: AwsEc2SecurityGroupUserIdGroupPair

Create a value of AwsEc2SecurityGroupUserIdGroupPair with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcPeeringConnectionId:AwsEc2SecurityGroupUserIdGroupPair', awsEc2SecurityGroupUserIdGroupPair_vpcPeeringConnectionId - The ID of the VPC peering connection, if applicable.

$sel:vpcId:AwsEc2SecurityGroupUserIdGroupPair', awsEc2SecurityGroupUserIdGroupPair_vpcId - The ID of the VPC for the referenced security group, if applicable.

$sel:userId:AwsEc2SecurityGroupUserIdGroupPair', awsEc2SecurityGroupUserIdGroupPair_userId - The ID of an Amazon Web Services account.

For a referenced security group in another VPC, the account ID of the referenced security group is returned in the response. If the referenced security group is deleted, this value is not returned.

EC2-Classic

Required when adding or removing rules that reference a security group in another VPC.

$sel:groupId:AwsEc2SecurityGroupUserIdGroupPair', awsEc2SecurityGroupUserIdGroupPair_groupId - The ID of the security group.

$sel:groupName:AwsEc2SecurityGroupUserIdGroupPair', awsEc2SecurityGroupUserIdGroupPair_groupName - The name of the security group.

$sel:peeringStatus:AwsEc2SecurityGroupUserIdGroupPair', awsEc2SecurityGroupUserIdGroupPair_peeringStatus - The status of a VPC peering connection, if applicable.

awsEc2SecurityGroupUserIdGroupPair_vpcPeeringConnectionId :: Lens' AwsEc2SecurityGroupUserIdGroupPair (Maybe Text)

The ID of the VPC peering connection, if applicable.

awsEc2SecurityGroupUserIdGroupPair_vpcId :: Lens' AwsEc2SecurityGroupUserIdGroupPair (Maybe Text)

The ID of the VPC for the referenced security group, if applicable.

awsEc2SecurityGroupUserIdGroupPair_userId :: Lens' AwsEc2SecurityGroupUserIdGroupPair (Maybe Text)

The ID of an Amazon Web Services account.

For a referenced security group in another VPC, the account ID of the referenced security group is returned in the response. If the referenced security group is deleted, this value is not returned.

EC2-Classic

Required when adding or removing rules that reference a security group in another VPC.

awsEc2SecurityGroupUserIdGroupPair_groupId :: Lens' AwsEc2SecurityGroupUserIdGroupPair (Maybe Text)

The ID of the security group.

awsEc2SecurityGroupUserIdGroupPair_groupName :: Lens' AwsEc2SecurityGroupUserIdGroupPair (Maybe Text)

The name of the security group.

awsEc2SecurityGroupUserIdGroupPair_peeringStatus :: Lens' AwsEc2SecurityGroupUserIdGroupPair (Maybe Text)

The status of a VPC peering connection, if applicable.