Amazonka.SecurityHub.Types.AwsEc2SubnetDetails

Description

Documentation

data AwsEc2SubnetDetails

Contains information about a subnet in Amazon EC2.

See: newAwsEc2SubnetDetails smart constructor.

Constructors

AwsEc2SubnetDetails'

Fields

state :: Maybe Text

The current state of the subnet.

ipv6CidrBlockAssociationSet :: Maybe [Ipv6CidrBlockAssociation]

The IPV6 CIDR blocks that are associated with the subnet.

availabilityZoneId :: Maybe Text

The identifier of the Availability Zone for the subnet.

availableIpAddressCount :: Maybe Int

The number of available IPV4 addresses in the subnet. Does not include addresses for stopped instances.

vpcId :: Maybe Text

The identifier of the VPC that contains the subnet.

assignIpv6AddressOnCreation :: Maybe Bool

Whether to assign an IPV6 address to a network interface that is created in this subnet.

subnetId :: Maybe Text

The identifier of the subnet.

subnetArn :: Maybe Text

The ARN of the subnet.

ownerId :: Maybe Text

The identifier of the Amazon Web Services account that owns the subnet.

availabilityZone :: Maybe Text

The Availability Zone for the subnet.

cidrBlock :: Maybe Text

The IPV4 CIDR block that is assigned to the subnet.

mapPublicIpOnLaunch :: Maybe Bool

Whether instances in this subnet receive a public IP address.

defaultForAz :: Maybe Bool

Whether this subnet is the default subnet for the Availability Zone.

newAwsEc2SubnetDetails :: AwsEc2SubnetDetails

Create a value of AwsEc2SubnetDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:AwsEc2SubnetDetails', awsEc2SubnetDetails_state - The current state of the subnet.

$sel:ipv6CidrBlockAssociationSet:AwsEc2SubnetDetails', awsEc2SubnetDetails_ipv6CidrBlockAssociationSet - The IPV6 CIDR blocks that are associated with the subnet.

$sel:availabilityZoneId:AwsEc2SubnetDetails', awsEc2SubnetDetails_availabilityZoneId - The identifier of the Availability Zone for the subnet.

$sel:availableIpAddressCount:AwsEc2SubnetDetails', awsEc2SubnetDetails_availableIpAddressCount - The number of available IPV4 addresses in the subnet. Does not include addresses for stopped instances.

$sel:vpcId:AwsEc2SubnetDetails', awsEc2SubnetDetails_vpcId - The identifier of the VPC that contains the subnet.

$sel:assignIpv6AddressOnCreation:AwsEc2SubnetDetails', awsEc2SubnetDetails_assignIpv6AddressOnCreation - Whether to assign an IPV6 address to a network interface that is created in this subnet.

$sel:subnetId:AwsEc2SubnetDetails', awsEc2SubnetDetails_subnetId - The identifier of the subnet.

$sel:subnetArn:AwsEc2SubnetDetails', awsEc2SubnetDetails_subnetArn - The ARN of the subnet.

$sel:ownerId:AwsEc2SubnetDetails', awsEc2SubnetDetails_ownerId - The identifier of the Amazon Web Services account that owns the subnet.

$sel:availabilityZone:AwsEc2SubnetDetails', awsEc2SubnetDetails_availabilityZone - The Availability Zone for the subnet.

$sel:cidrBlock:AwsEc2SubnetDetails', awsEc2SubnetDetails_cidrBlock - The IPV4 CIDR block that is assigned to the subnet.

$sel:mapPublicIpOnLaunch:AwsEc2SubnetDetails', awsEc2SubnetDetails_mapPublicIpOnLaunch - Whether instances in this subnet receive a public IP address.

$sel:defaultForAz:AwsEc2SubnetDetails', awsEc2SubnetDetails_defaultForAz - Whether this subnet is the default subnet for the Availability Zone.

awsEc2SubnetDetails_state :: Lens' AwsEc2SubnetDetails (Maybe Text)

The current state of the subnet.

awsEc2SubnetDetails_ipv6CidrBlockAssociationSet :: Lens' AwsEc2SubnetDetails (Maybe [Ipv6CidrBlockAssociation])

The IPV6 CIDR blocks that are associated with the subnet.

awsEc2SubnetDetails_availabilityZoneId :: Lens' AwsEc2SubnetDetails (Maybe Text)

The identifier of the Availability Zone for the subnet.

awsEc2SubnetDetails_availableIpAddressCount :: Lens' AwsEc2SubnetDetails (Maybe Int)

The number of available IPV4 addresses in the subnet. Does not include addresses for stopped instances.

awsEc2SubnetDetails_vpcId :: Lens' AwsEc2SubnetDetails (Maybe Text)

The identifier of the VPC that contains the subnet.

awsEc2SubnetDetails_assignIpv6AddressOnCreation :: Lens' AwsEc2SubnetDetails (Maybe Bool)

Whether to assign an IPV6 address to a network interface that is created in this subnet.

awsEc2SubnetDetails_subnetId :: Lens' AwsEc2SubnetDetails (Maybe Text)

The identifier of the subnet.

awsEc2SubnetDetails_subnetArn :: Lens' AwsEc2SubnetDetails (Maybe Text)

The ARN of the subnet.

awsEc2SubnetDetails_ownerId :: Lens' AwsEc2SubnetDetails (Maybe Text)

The identifier of the Amazon Web Services account that owns the subnet.

awsEc2SubnetDetails_availabilityZone :: Lens' AwsEc2SubnetDetails (Maybe Text)

The Availability Zone for the subnet.

awsEc2SubnetDetails_cidrBlock :: Lens' AwsEc2SubnetDetails (Maybe Text)

The IPV4 CIDR block that is assigned to the subnet.

awsEc2SubnetDetails_mapPublicIpOnLaunch :: Lens' AwsEc2SubnetDetails (Maybe Bool)

Whether instances in this subnet receive a public IP address.

awsEc2SubnetDetails_defaultForAz :: Lens' AwsEc2SubnetDetails (Maybe Bool)

Whether this subnet is the default subnet for the Availability Zone.