Amazonka.SecurityHub.Types.AwsEc2VolumeDetails

Description

Documentation

data AwsEc2VolumeDetails

Details about an EC2 volume.

See: newAwsEc2VolumeDetails smart constructor.

Constructors

AwsEc2VolumeDetails'

Fields

status :: Maybe Text

The volume state.

attachments :: Maybe [AwsEc2VolumeAttachment]

The volume attachments.

size :: Maybe Int

The size of the volume, in GiBs.

encrypted :: Maybe Bool

Whether the volume is encrypted.

kmsKeyId :: Maybe Text

The ARN of the KMS key that was used to protect the volume encryption key for the volume.

createTime :: Maybe Text

Indicates when the volume was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

snapshotId :: Maybe Text

The snapshot from which the volume was created.

newAwsEc2VolumeDetails :: AwsEc2VolumeDetails

Create a value of AwsEc2VolumeDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AwsEc2VolumeDetails', awsEc2VolumeDetails_status - The volume state.

$sel:attachments:AwsEc2VolumeDetails', awsEc2VolumeDetails_attachments - The volume attachments.

$sel:size:AwsEc2VolumeDetails', awsEc2VolumeDetails_size - The size of the volume, in GiBs.

$sel:encrypted:AwsEc2VolumeDetails', awsEc2VolumeDetails_encrypted - Whether the volume is encrypted.

$sel:kmsKeyId:AwsEc2VolumeDetails', awsEc2VolumeDetails_kmsKeyId - The ARN of the KMS key that was used to protect the volume encryption key for the volume.

$sel:createTime:AwsEc2VolumeDetails', awsEc2VolumeDetails_createTime - Indicates when the volume was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:snapshotId:AwsEc2VolumeDetails', awsEc2VolumeDetails_snapshotId - The snapshot from which the volume was created.

awsEc2VolumeDetails_status :: Lens' AwsEc2VolumeDetails (Maybe Text)

The volume state.

awsEc2VolumeDetails_attachments :: Lens' AwsEc2VolumeDetails (Maybe [AwsEc2VolumeAttachment])

The volume attachments.

awsEc2VolumeDetails_size :: Lens' AwsEc2VolumeDetails (Maybe Int)

The size of the volume, in GiBs.

awsEc2VolumeDetails_encrypted :: Lens' AwsEc2VolumeDetails (Maybe Bool)

Whether the volume is encrypted.

awsEc2VolumeDetails_kmsKeyId :: Lens' AwsEc2VolumeDetails (Maybe Text)

The ARN of the KMS key that was used to protect the volume encryption key for the volume.

awsEc2VolumeDetails_createTime :: Lens' AwsEc2VolumeDetails (Maybe Text)

Indicates when the volume was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsEc2VolumeDetails_snapshotId :: Lens' AwsEc2VolumeDetails (Maybe Text)

The snapshot from which the volume was created.