Amazonka.SecurityHub.Types.AwsEc2VpnConnectionDetails

Description

Documentation

data AwsEc2VpnConnectionDetails

Details about an Amazon EC2 VPN connection.

See: newAwsEc2VpnConnectionDetails smart constructor.

Constructors

AwsEc2VpnConnectionDetails'

Fields

customerGatewayConfiguration :: Maybe Text

The configuration information for the VPN connection's customer gateway, in the native XML format.

state :: Maybe Text

The current state of the VPN connection.

routes :: Maybe [AwsEc2VpnConnectionRoutesDetails]

The static routes that are associated with the VPN connection.

vpnGatewayId :: Maybe Text

The identifier of the virtual private gateway that is at the Amazon Web Services side of the VPN connection.

category :: Maybe Text

The category of the VPN connection. VPN indicates an Amazon Web Services VPN connection. VPN-Classic indicates an Amazon Web Services Classic VPN connection.

customerGatewayId :: Maybe Text

The identifier of the customer gateway that is at your end of the VPN connection.

transitGatewayId :: Maybe Text

The identifier of the transit gateway that is associated with the VPN connection.

type' :: Maybe Text

The type of VPN connection.

options :: Maybe AwsEc2VpnConnectionOptionsDetails

The VPN connection options.

vpnConnectionId :: Maybe Text

The identifier of the VPN connection.

vgwTelemetry :: Maybe [AwsEc2VpnConnectionVgwTelemetryDetails]

Information about the VPN tunnel.

newAwsEc2VpnConnectionDetails :: AwsEc2VpnConnectionDetails

Create a value of AwsEc2VpnConnectionDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customerGatewayConfiguration:AwsEc2VpnConnectionDetails', awsEc2VpnConnectionDetails_customerGatewayConfiguration - The configuration information for the VPN connection's customer gateway, in the native XML format.

$sel:state:AwsEc2VpnConnectionDetails', awsEc2VpnConnectionDetails_state - The current state of the VPN connection.

$sel:routes:AwsEc2VpnConnectionDetails', awsEc2VpnConnectionDetails_routes - The static routes that are associated with the VPN connection.

$sel:vpnGatewayId:AwsEc2VpnConnectionDetails', awsEc2VpnConnectionDetails_vpnGatewayId - The identifier of the virtual private gateway that is at the Amazon Web Services side of the VPN connection.

$sel:category:AwsEc2VpnConnectionDetails', awsEc2VpnConnectionDetails_category - The category of the VPN connection. VPN indicates an Amazon Web Services VPN connection. VPN-Classic indicates an Amazon Web Services Classic VPN connection.

$sel:customerGatewayId:AwsEc2VpnConnectionDetails', awsEc2VpnConnectionDetails_customerGatewayId - The identifier of the customer gateway that is at your end of the VPN connection.

$sel:transitGatewayId:AwsEc2VpnConnectionDetails', awsEc2VpnConnectionDetails_transitGatewayId - The identifier of the transit gateway that is associated with the VPN connection.

$sel:type':AwsEc2VpnConnectionDetails', awsEc2VpnConnectionDetails_type - The type of VPN connection.

$sel:options:AwsEc2VpnConnectionDetails', awsEc2VpnConnectionDetails_options - The VPN connection options.

$sel:vpnConnectionId:AwsEc2VpnConnectionDetails', awsEc2VpnConnectionDetails_vpnConnectionId - The identifier of the VPN connection.

$sel:vgwTelemetry:AwsEc2VpnConnectionDetails', awsEc2VpnConnectionDetails_vgwTelemetry - Information about the VPN tunnel.

awsEc2VpnConnectionDetails_customerGatewayConfiguration :: Lens' AwsEc2VpnConnectionDetails (Maybe Text)

The configuration information for the VPN connection's customer gateway, in the native XML format.

awsEc2VpnConnectionDetails_state :: Lens' AwsEc2VpnConnectionDetails (Maybe Text)

The current state of the VPN connection.

awsEc2VpnConnectionDetails_routes :: Lens' AwsEc2VpnConnectionDetails (Maybe [AwsEc2VpnConnectionRoutesDetails])

The static routes that are associated with the VPN connection.

awsEc2VpnConnectionDetails_vpnGatewayId :: Lens' AwsEc2VpnConnectionDetails (Maybe Text)

The identifier of the virtual private gateway that is at the Amazon Web Services side of the VPN connection.

awsEc2VpnConnectionDetails_category :: Lens' AwsEc2VpnConnectionDetails (Maybe Text)

The category of the VPN connection. VPN indicates an Amazon Web Services VPN connection. VPN-Classic indicates an Amazon Web Services Classic VPN connection.

awsEc2VpnConnectionDetails_customerGatewayId :: Lens' AwsEc2VpnConnectionDetails (Maybe Text)

The identifier of the customer gateway that is at your end of the VPN connection.

awsEc2VpnConnectionDetails_transitGatewayId :: Lens' AwsEc2VpnConnectionDetails (Maybe Text)

The identifier of the transit gateway that is associated with the VPN connection.

awsEc2VpnConnectionDetails_type :: Lens' AwsEc2VpnConnectionDetails (Maybe Text)

The type of VPN connection.

awsEc2VpnConnectionDetails_options :: Lens' AwsEc2VpnConnectionDetails (Maybe AwsEc2VpnConnectionOptionsDetails)

The VPN connection options.

awsEc2VpnConnectionDetails_vpnConnectionId :: Lens' AwsEc2VpnConnectionDetails (Maybe Text)

The identifier of the VPN connection.

awsEc2VpnConnectionDetails_vgwTelemetry :: Lens' AwsEc2VpnConnectionDetails (Maybe [AwsEc2VpnConnectionVgwTelemetryDetails])

Information about the VPN tunnel.