Amazonka.SecurityHub.Types.AwsEc2VpnConnectionRoutesDetails

Description

Documentation

data AwsEc2VpnConnectionRoutesDetails

A static routes associated with the VPN connection.

See: newAwsEc2VpnConnectionRoutesDetails smart constructor.

Constructors

AwsEc2VpnConnectionRoutesDetails'

Fields

state :: Maybe Text

The current state of the static route.

destinationCidrBlock :: Maybe Text

The CIDR block associated with the local subnet of the customer data center.

newAwsEc2VpnConnectionRoutesDetails :: AwsEc2VpnConnectionRoutesDetails

Create a value of AwsEc2VpnConnectionRoutesDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:AwsEc2VpnConnectionRoutesDetails', awsEc2VpnConnectionRoutesDetails_state - The current state of the static route.

$sel:destinationCidrBlock:AwsEc2VpnConnectionRoutesDetails', awsEc2VpnConnectionRoutesDetails_destinationCidrBlock - The CIDR block associated with the local subnet of the customer data center.

awsEc2VpnConnectionRoutesDetails_state :: Lens' AwsEc2VpnConnectionRoutesDetails (Maybe Text)

The current state of the static route.

awsEc2VpnConnectionRoutesDetails_destinationCidrBlock :: Lens' AwsEc2VpnConnectionRoutesDetails (Maybe Text)

The CIDR block associated with the local subnet of the customer data center.