Amazonka.SecurityHub.Types.AwsEcsServiceDeploymentConfigurationDetails

Description

Documentation

data AwsEcsServiceDeploymentConfigurationDetails

Optional deployment parameters for the service.

See: newAwsEcsServiceDeploymentConfigurationDetails smart constructor.

Constructors

AwsEcsServiceDeploymentConfigurationDetails'

Fields

minimumHealthyPercent :: Maybe Int

For a service that uses the rolling update (ECS) deployment type, the minimum number of tasks in a service that must remain in the RUNNING state during a deployment, and while any container instances are in the DRAINING state if the service contains tasks using the EC2 launch type. Expressed as a percentage of the desired number of tasks. The default value is 100%.

For a service that uses the blue/green (CODE_DEPLOY) or EXTERNAL deployment types and tasks that use the EC2 launch type, the minimum number of the tasks in the service that remain in the RUNNING state while the container instances are in the DRAINING state.

For the Fargate launch type, the minimum healthy percent value is not used.

maximumPercent :: Maybe Int

For a service that uses the rolling update (ECS) deployment type, the maximum number of tasks in a service that are allowed in the RUNNING or PENDING state during a deployment, and for tasks that use the EC2 launch type, when any container instances are in the DRAINING state. Provided as a percentage of the desired number of tasks. The default value is 200%.

For a service that uses the blue/green (CODE_DEPLOY) or EXTERNAL deployment types, and tasks that use the EC2 launch type, the maximum number of tasks in the service that remain in the RUNNING state while the container instances are in the DRAINING state.

For the Fargate launch type, the maximum percent value is not used.

deploymentCircuitBreaker :: Maybe AwsEcsServiceDeploymentConfigurationDeploymentCircuitBreakerDetails

Determines whether a service deployment fails if a service cannot reach a steady state.

newAwsEcsServiceDeploymentConfigurationDetails :: AwsEcsServiceDeploymentConfigurationDetails

Create a value of AwsEcsServiceDeploymentConfigurationDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minimumHealthyPercent:AwsEcsServiceDeploymentConfigurationDetails', awsEcsServiceDeploymentConfigurationDetails_minimumHealthyPercent - For a service that uses the rolling update (ECS) deployment type, the minimum number of tasks in a service that must remain in the RUNNING state during a deployment, and while any container instances are in the DRAINING state if the service contains tasks using the EC2 launch type. Expressed as a percentage of the desired number of tasks. The default value is 100%.

For a service that uses the blue/green (CODE_DEPLOY) or EXTERNAL deployment types and tasks that use the EC2 launch type, the minimum number of the tasks in the service that remain in the RUNNING state while the container instances are in the DRAINING state.

For the Fargate launch type, the minimum healthy percent value is not used.

$sel:maximumPercent:AwsEcsServiceDeploymentConfigurationDetails', awsEcsServiceDeploymentConfigurationDetails_maximumPercent - For a service that uses the rolling update (ECS) deployment type, the maximum number of tasks in a service that are allowed in the RUNNING or PENDING state during a deployment, and for tasks that use the EC2 launch type, when any container instances are in the DRAINING state. Provided as a percentage of the desired number of tasks. The default value is 200%.

For a service that uses the blue/green (CODE_DEPLOY) or EXTERNAL deployment types, and tasks that use the EC2 launch type, the maximum number of tasks in the service that remain in the RUNNING state while the container instances are in the DRAINING state.

For the Fargate launch type, the maximum percent value is not used.

$sel:deploymentCircuitBreaker:AwsEcsServiceDeploymentConfigurationDetails', awsEcsServiceDeploymentConfigurationDetails_deploymentCircuitBreaker - Determines whether a service deployment fails if a service cannot reach a steady state.

awsEcsServiceDeploymentConfigurationDetails_minimumHealthyPercent :: Lens' AwsEcsServiceDeploymentConfigurationDetails (Maybe Int)

For a service that uses the rolling update (ECS) deployment type, the minimum number of tasks in a service that must remain in the RUNNING state during a deployment, and while any container instances are in the DRAINING state if the service contains tasks using the EC2 launch type. Expressed as a percentage of the desired number of tasks. The default value is 100%.

For a service that uses the blue/green (CODE_DEPLOY) or EXTERNAL deployment types and tasks that use the EC2 launch type, the minimum number of the tasks in the service that remain in the RUNNING state while the container instances are in the DRAINING state.

For the Fargate launch type, the minimum healthy percent value is not used.

awsEcsServiceDeploymentConfigurationDetails_maximumPercent :: Lens' AwsEcsServiceDeploymentConfigurationDetails (Maybe Int)

For a service that uses the rolling update (ECS) deployment type, the maximum number of tasks in a service that are allowed in the RUNNING or PENDING state during a deployment, and for tasks that use the EC2 launch type, when any container instances are in the DRAINING state. Provided as a percentage of the desired number of tasks. The default value is 200%.

For a service that uses the blue/green (CODE_DEPLOY) or EXTERNAL deployment types, and tasks that use the EC2 launch type, the maximum number of tasks in the service that remain in the RUNNING state while the container instances are in the DRAINING state.

For the Fargate launch type, the maximum percent value is not used.

awsEcsServiceDeploymentConfigurationDetails_deploymentCircuitBreaker :: Lens' AwsEcsServiceDeploymentConfigurationDetails (Maybe AwsEcsServiceDeploymentConfigurationDeploymentCircuitBreakerDetails)

Determines whether a service deployment fails if a service cannot reach a steady state.