Amazonka.SecurityHub.Types.AwsEcsServiceDetails

Description

Documentation

data AwsEcsServiceDetails

Provides details about a service within an ECS cluster.

See: newAwsEcsServiceDetails smart constructor.

Constructors

AwsEcsServiceDetails'

Fields

placementStrategies :: Maybe [AwsEcsServicePlacementStrategiesDetails]

Information about how tasks for the service are placed.

cluster :: Maybe Text

The ARN of the cluster that hosts the service.

propagateTags :: Maybe Text

Indicates whether to propagate the tags from the task definition to the task or from the service to the task. If no value is provided, then tags are not propagated.

Valid values: TASK_DEFINITION | SERVICE

platformVersion :: Maybe Text

The platform version on which to run the service. Only specified for tasks that are hosted on Fargate. If a platform version is not specified, the LATEST platform version is used by default.

enableEcsManagedTags :: Maybe Bool

Whether to enable Amazon ECS managed tags for the tasks in the service.

desiredCount :: Maybe Int

The number of instantiations of the task definition to run on the service.

loadBalancers :: Maybe [AwsEcsServiceLoadBalancersDetails]

Information about the load balancers that the service uses.

role' :: Maybe Text

The ARN of the IAM role that is associated with the service. The role allows the Amazon ECS container agent to register container instances with an Elastic Load Balancing load balancer.

name :: Maybe Text

The name of the service.

placementConstraints :: Maybe [AwsEcsServicePlacementConstraintsDetails]

The placement constraints for the tasks in the service.

serviceName :: Maybe Text

The name of the service.

The name can contain up to 255 characters. It can use letters, numbers, underscores, and hyphens.

deploymentController :: Maybe AwsEcsServiceDeploymentControllerDetails

Contains the deployment controller type that the service uses.

launchType :: Maybe Text

The launch type that the service uses.

Valid values: EC2 | FARGATE | EXTERNAL

taskDefinition :: Maybe Text

The task definition to use for tasks in the service.

serviceArn :: Maybe Text

The ARN of the service.

schedulingStrategy :: Maybe Text

The scheduling strategy to use for the service.

The REPLICA scheduling strategy places and maintains the desired number of tasks across the cluster. By default, the service scheduler spreads tasks across Availability Zones. Task placement strategies and constraints are used to customize task placement decisions.

The DAEMON scheduling strategy deploys exactly one task on each active container instance that meets all of the task placement constraints that are specified in the cluster. The service scheduler also evaluates the task placement constraints for running tasks and stops tasks that do not meet the placement constraints.

Valid values: REPLICA | DAEMON

capacityProviderStrategy :: Maybe [AwsEcsServiceCapacityProviderStrategyDetails]

The capacity provider strategy that the service uses.

serviceRegistries :: Maybe [AwsEcsServiceServiceRegistriesDetails]

Information about the service discovery registries to assign to the service.

healthCheckGracePeriodSeconds :: Maybe Int

After a task starts, the amount of time in seconds that the Amazon ECS service scheduler ignores unhealthy Elastic Load Balancing target health checks.

networkConfiguration :: Maybe AwsEcsServiceNetworkConfigurationDetails

For tasks that use the awsvpc networking mode, the VPC subnet and security group configuration.

deploymentConfiguration :: Maybe AwsEcsServiceDeploymentConfigurationDetails

Deployment parameters for the service. Includes the number of tasks that run and the order in which to start and stop tasks.

enableExecuteCommand :: Maybe Bool

Whether the execute command functionality is enabled for the service.

newAwsEcsServiceDetails :: AwsEcsServiceDetails

Create a value of AwsEcsServiceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:placementStrategies:AwsEcsServiceDetails', awsEcsServiceDetails_placementStrategies - Information about how tasks for the service are placed.

$sel:cluster:AwsEcsServiceDetails', awsEcsServiceDetails_cluster - The ARN of the cluster that hosts the service.

$sel:propagateTags:AwsEcsServiceDetails', awsEcsServiceDetails_propagateTags - Indicates whether to propagate the tags from the task definition to the task or from the service to the task. If no value is provided, then tags are not propagated.

Valid values: TASK_DEFINITION | SERVICE

$sel:platformVersion:AwsEcsServiceDetails', awsEcsServiceDetails_platformVersion - The platform version on which to run the service. Only specified for tasks that are hosted on Fargate. If a platform version is not specified, the LATEST platform version is used by default.

$sel:enableEcsManagedTags:AwsEcsServiceDetails', awsEcsServiceDetails_enableEcsManagedTags - Whether to enable Amazon ECS managed tags for the tasks in the service.

$sel:desiredCount:AwsEcsServiceDetails', awsEcsServiceDetails_desiredCount - The number of instantiations of the task definition to run on the service.

$sel:loadBalancers:AwsEcsServiceDetails', awsEcsServiceDetails_loadBalancers - Information about the load balancers that the service uses.

$sel:role':AwsEcsServiceDetails', awsEcsServiceDetails_role - The ARN of the IAM role that is associated with the service. The role allows the Amazon ECS container agent to register container instances with an Elastic Load Balancing load balancer.

$sel:name:AwsEcsServiceDetails', awsEcsServiceDetails_name - The name of the service.

$sel:placementConstraints:AwsEcsServiceDetails', awsEcsServiceDetails_placementConstraints - The placement constraints for the tasks in the service.

$sel:serviceName:AwsEcsServiceDetails', awsEcsServiceDetails_serviceName - The name of the service.

The name can contain up to 255 characters. It can use letters, numbers, underscores, and hyphens.

$sel:deploymentController:AwsEcsServiceDetails', awsEcsServiceDetails_deploymentController - Contains the deployment controller type that the service uses.

$sel:launchType:AwsEcsServiceDetails', awsEcsServiceDetails_launchType - The launch type that the service uses.

Valid values: EC2 | FARGATE | EXTERNAL

$sel:taskDefinition:AwsEcsServiceDetails', awsEcsServiceDetails_taskDefinition - The task definition to use for tasks in the service.

$sel:serviceArn:AwsEcsServiceDetails', awsEcsServiceDetails_serviceArn - The ARN of the service.

$sel:schedulingStrategy:AwsEcsServiceDetails', awsEcsServiceDetails_schedulingStrategy - The scheduling strategy to use for the service.

The REPLICA scheduling strategy places and maintains the desired number of tasks across the cluster. By default, the service scheduler spreads tasks across Availability Zones. Task placement strategies and constraints are used to customize task placement decisions.

The DAEMON scheduling strategy deploys exactly one task on each active container instance that meets all of the task placement constraints that are specified in the cluster. The service scheduler also evaluates the task placement constraints for running tasks and stops tasks that do not meet the placement constraints.

Valid values: REPLICA | DAEMON

$sel:capacityProviderStrategy:AwsEcsServiceDetails', awsEcsServiceDetails_capacityProviderStrategy - The capacity provider strategy that the service uses.

$sel:serviceRegistries:AwsEcsServiceDetails', awsEcsServiceDetails_serviceRegistries - Information about the service discovery registries to assign to the service.

$sel:healthCheckGracePeriodSeconds:AwsEcsServiceDetails', awsEcsServiceDetails_healthCheckGracePeriodSeconds - After a task starts, the amount of time in seconds that the Amazon ECS service scheduler ignores unhealthy Elastic Load Balancing target health checks.

$sel:networkConfiguration:AwsEcsServiceDetails', awsEcsServiceDetails_networkConfiguration - For tasks that use the awsvpc networking mode, the VPC subnet and security group configuration.

$sel:deploymentConfiguration:AwsEcsServiceDetails', awsEcsServiceDetails_deploymentConfiguration - Deployment parameters for the service. Includes the number of tasks that run and the order in which to start and stop tasks.

$sel:enableExecuteCommand:AwsEcsServiceDetails', awsEcsServiceDetails_enableExecuteCommand - Whether the execute command functionality is enabled for the service.

awsEcsServiceDetails_placementStrategies :: Lens' AwsEcsServiceDetails (Maybe [AwsEcsServicePlacementStrategiesDetails])

Information about how tasks for the service are placed.

awsEcsServiceDetails_cluster :: Lens' AwsEcsServiceDetails (Maybe Text)

The ARN of the cluster that hosts the service.

awsEcsServiceDetails_propagateTags :: Lens' AwsEcsServiceDetails (Maybe Text)

Indicates whether to propagate the tags from the task definition to the task or from the service to the task. If no value is provided, then tags are not propagated.

Valid values: TASK_DEFINITION | SERVICE

awsEcsServiceDetails_platformVersion :: Lens' AwsEcsServiceDetails (Maybe Text)

The platform version on which to run the service. Only specified for tasks that are hosted on Fargate. If a platform version is not specified, the LATEST platform version is used by default.

awsEcsServiceDetails_enableEcsManagedTags :: Lens' AwsEcsServiceDetails (Maybe Bool)

Whether to enable Amazon ECS managed tags for the tasks in the service.

awsEcsServiceDetails_desiredCount :: Lens' AwsEcsServiceDetails (Maybe Int)

The number of instantiations of the task definition to run on the service.

awsEcsServiceDetails_loadBalancers :: Lens' AwsEcsServiceDetails (Maybe [AwsEcsServiceLoadBalancersDetails])

Information about the load balancers that the service uses.

awsEcsServiceDetails_role :: Lens' AwsEcsServiceDetails (Maybe Text)

The ARN of the IAM role that is associated with the service. The role allows the Amazon ECS container agent to register container instances with an Elastic Load Balancing load balancer.

awsEcsServiceDetails_name :: Lens' AwsEcsServiceDetails (Maybe Text)

The name of the service.

awsEcsServiceDetails_placementConstraints :: Lens' AwsEcsServiceDetails (Maybe [AwsEcsServicePlacementConstraintsDetails])

The placement constraints for the tasks in the service.

awsEcsServiceDetails_serviceName :: Lens' AwsEcsServiceDetails (Maybe Text)

The name of the service.

The name can contain up to 255 characters. It can use letters, numbers, underscores, and hyphens.

awsEcsServiceDetails_deploymentController :: Lens' AwsEcsServiceDetails (Maybe AwsEcsServiceDeploymentControllerDetails)

Contains the deployment controller type that the service uses.

awsEcsServiceDetails_launchType :: Lens' AwsEcsServiceDetails (Maybe Text)

The launch type that the service uses.

Valid values: EC2 | FARGATE | EXTERNAL

awsEcsServiceDetails_taskDefinition :: Lens' AwsEcsServiceDetails (Maybe Text)

The task definition to use for tasks in the service.

awsEcsServiceDetails_serviceArn :: Lens' AwsEcsServiceDetails (Maybe Text)

The ARN of the service.

awsEcsServiceDetails_schedulingStrategy :: Lens' AwsEcsServiceDetails (Maybe Text)

The scheduling strategy to use for the service.

The REPLICA scheduling strategy places and maintains the desired number of tasks across the cluster. By default, the service scheduler spreads tasks across Availability Zones. Task placement strategies and constraints are used to customize task placement decisions.

The DAEMON scheduling strategy deploys exactly one task on each active container instance that meets all of the task placement constraints that are specified in the cluster. The service scheduler also evaluates the task placement constraints for running tasks and stops tasks that do not meet the placement constraints.

Valid values: REPLICA | DAEMON

awsEcsServiceDetails_capacityProviderStrategy :: Lens' AwsEcsServiceDetails (Maybe [AwsEcsServiceCapacityProviderStrategyDetails])

The capacity provider strategy that the service uses.

awsEcsServiceDetails_serviceRegistries :: Lens' AwsEcsServiceDetails (Maybe [AwsEcsServiceServiceRegistriesDetails])

Information about the service discovery registries to assign to the service.

awsEcsServiceDetails_healthCheckGracePeriodSeconds :: Lens' AwsEcsServiceDetails (Maybe Int)

After a task starts, the amount of time in seconds that the Amazon ECS service scheduler ignores unhealthy Elastic Load Balancing target health checks.

awsEcsServiceDetails_networkConfiguration :: Lens' AwsEcsServiceDetails (Maybe AwsEcsServiceNetworkConfigurationDetails)

For tasks that use the awsvpc networking mode, the VPC subnet and security group configuration.

awsEcsServiceDetails_deploymentConfiguration :: Lens' AwsEcsServiceDetails (Maybe AwsEcsServiceDeploymentConfigurationDetails)

Deployment parameters for the service. Includes the number of tasks that run and the order in which to start and stop tasks.

awsEcsServiceDetails_enableExecuteCommand :: Lens' AwsEcsServiceDetails (Maybe Bool)

Whether the execute command functionality is enabled for the service.