Amazonka.SecurityHub.Types.AwsEcsTaskDefinitionContainerDefinitionsDetails

Description

Documentation

data AwsEcsTaskDefinitionContainerDefinitionsDetails

A container definition that describes a container in the task.

See: newAwsEcsTaskDefinitionContainerDefinitionsDetails smart constructor.

Constructors

AwsEcsTaskDefinitionContainerDefinitionsDetails'

Fields

command :: Maybe [Text]

The command that is passed to the container.

hostname :: Maybe Text

The hostname to use for the container.

image :: Maybe Text

The image used to start the container.

repositoryCredentials :: Maybe AwsEcsTaskDefinitionContainerDefinitionsRepositoryCredentialsDetails

The private repository authentication credentials to use.

dockerSecurityOptions :: Maybe [Text]

A list of strings to provide custom labels for SELinux and AppArmor multi-level security systems.

healthCheck :: Maybe AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails

The container health check command and associated configuration parameters for the container.

disableNetworking :: Maybe Bool

Whether to disable networking within the container.

volumesFrom :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails]

Data volumes to mount from another container.

environment :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsEnvironmentDetails]

The environment variables to pass to a container.

secrets :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsSecretsDetails]

The secrets to pass to the container.

entryPoint :: Maybe [Text]

The entry point that is passed to the container.

workingDirectory :: Maybe Text

The working directory in which to run commands inside the container.

environmentFiles :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsEnvironmentFilesDetails]

A list of files containing the environment variables to pass to a container.

stopTimeout :: Maybe Int

The number of seconds to wait before the container is stopped if it doesn't shut down normally on its own.

privileged :: Maybe Bool

Whether the container is given elevated privileges on the host container instance. The elevated privileges are similar to the root user.

ulimits :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsUlimitsDetails]

A list of ulimits to set in the container.

portMappings :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsPortMappingsDetails]

The list of port mappings for the container.

resourceRequirements :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsResourceRequirementsDetails]

The type and amount of a resource to assign to a container. The only supported resource is a GPU.

memory :: Maybe Int

The amount (in MiB) of memory to present to the container. If the container attempts to exceed the memory specified here, the container is shut down. The total amount of memory reserved for all containers within a task must be lower than the task memory value, if one is specified.

extraHosts :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsExtraHostsDetails]

A list of hostnames and IP address mappings to append to the /etc/hosts file on the container.

dockerLabels :: Maybe (HashMap Text Text)

A key-value map of labels to add to the container.

user :: Maybe Text

The user to use inside the container.

The value can use one of the following formats.

 user

user : group

 uid

uid : gid

user : gid

uid : group

systemControls :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsSystemControlsDetails]

A list of namespaced kernel parameters to set in the container.

linuxParameters :: Maybe AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails

Linux-specific modifications that are applied to the container, such as Linux kernel capabilities.

logConfiguration :: Maybe AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails

The log configuration specification for the container.

firelensConfiguration :: Maybe AwsEcsTaskDefinitionContainerDefinitionsFirelensConfigurationDetails

The FireLens configuration for the container. Specifies and configures a log router for container logs.

dnsSearchDomains :: Maybe [Text]

A list of DNS search domains that are presented to the container.

pseudoTerminal :: Maybe Bool

Whether to allocate a TTY to the container.

dependsOn :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsDependsOnDetails]

The dependencies that are defined for container startup and shutdown.

dnsServers :: Maybe [Text]

A list of DNS servers that are presented to the container.

name :: Maybe Text

The name of the container.

mountPoints :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsMountPointsDetails]

The mount points for the data volumes in the container.

interactive :: Maybe Bool

If set to true, then containerized applications can be deployed that require stdin or a tty to be allocated.

startTimeout :: Maybe Int

The number of seconds to wait before giving up on resolving dependencies for a container.

readonlyRootFilesystem :: Maybe Bool

Whether the container is given read-only access to its root file system.

links :: Maybe [Text]

A list of links for the container in the form container_name:alias . Allows containers to communicate with each other without the need for port mappings.

cpu :: Maybe Int

The number of CPU units reserved for the container.

essential :: Maybe Bool

Whether the container is essential. All tasks must have at least one essential container.

memoryReservation :: Maybe Int

The soft limit (in MiB) of memory to reserve for the container.

newAwsEcsTaskDefinitionContainerDefinitionsDetails :: AwsEcsTaskDefinitionContainerDefinitionsDetails

Create a value of AwsEcsTaskDefinitionContainerDefinitionsDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:command:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_command - The command that is passed to the container.

$sel:hostname:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_hostname - The hostname to use for the container.

$sel:image:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_image - The image used to start the container.

$sel:repositoryCredentials:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_repositoryCredentials - The private repository authentication credentials to use.

$sel:dockerSecurityOptions:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_dockerSecurityOptions - A list of strings to provide custom labels for SELinux and AppArmor multi-level security systems.

$sel:healthCheck:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_healthCheck - The container health check command and associated configuration parameters for the container.

$sel:disableNetworking:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_disableNetworking - Whether to disable networking within the container.

$sel:volumesFrom:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_volumesFrom - Data volumes to mount from another container.

$sel:environment:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_environment - The environment variables to pass to a container.

$sel:secrets:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_secrets - The secrets to pass to the container.

$sel:entryPoint:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_entryPoint - The entry point that is passed to the container.

$sel:workingDirectory:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_workingDirectory - The working directory in which to run commands inside the container.

$sel:environmentFiles:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_environmentFiles - A list of files containing the environment variables to pass to a container.

$sel:stopTimeout:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_stopTimeout - The number of seconds to wait before the container is stopped if it doesn't shut down normally on its own.

$sel:privileged:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_privileged - Whether the container is given elevated privileges on the host container instance. The elevated privileges are similar to the root user.

$sel:ulimits:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_ulimits - A list of ulimits to set in the container.

$sel:portMappings:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_portMappings - The list of port mappings for the container.

$sel:resourceRequirements:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_resourceRequirements - The type and amount of a resource to assign to a container. The only supported resource is a GPU.

$sel:memory:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_memory - The amount (in MiB) of memory to present to the container. If the container attempts to exceed the memory specified here, the container is shut down. The total amount of memory reserved for all containers within a task must be lower than the task memory value, if one is specified.

$sel:extraHosts:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_extraHosts - A list of hostnames and IP address mappings to append to the /etc/hosts file on the container.

$sel:dockerLabels:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_dockerLabels - A key-value map of labels to add to the container.

$sel:user:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_user - The user to use inside the container.

The value can use one of the following formats.

 user

user : group

 uid

uid : gid

user : gid

uid : group

$sel:systemControls:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_systemControls - A list of namespaced kernel parameters to set in the container.

$sel:linuxParameters:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_linuxParameters - Linux-specific modifications that are applied to the container, such as Linux kernel capabilities.

$sel:logConfiguration:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_logConfiguration - The log configuration specification for the container.

$sel:firelensConfiguration:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_firelensConfiguration - The FireLens configuration for the container. Specifies and configures a log router for container logs.

$sel:dnsSearchDomains:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_dnsSearchDomains - A list of DNS search domains that are presented to the container.

$sel:pseudoTerminal:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_pseudoTerminal - Whether to allocate a TTY to the container.

$sel:dependsOn:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_dependsOn - The dependencies that are defined for container startup and shutdown.

$sel:dnsServers:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_dnsServers - A list of DNS servers that are presented to the container.

$sel:name:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_name - The name of the container.

$sel:mountPoints:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_mountPoints - The mount points for the data volumes in the container.

$sel:interactive:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_interactive - If set to true, then containerized applications can be deployed that require stdin or a tty to be allocated.

$sel:startTimeout:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_startTimeout - The number of seconds to wait before giving up on resolving dependencies for a container.

$sel:readonlyRootFilesystem:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_readonlyRootFilesystem - Whether the container is given read-only access to its root file system.

$sel:links:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_links - A list of links for the container in the form container_name:alias . Allows containers to communicate with each other without the need for port mappings.

$sel:cpu:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_cpu - The number of CPU units reserved for the container.

$sel:essential:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_essential - Whether the container is essential. All tasks must have at least one essential container.

$sel:memoryReservation:AwsEcsTaskDefinitionContainerDefinitionsDetails', awsEcsTaskDefinitionContainerDefinitionsDetails_memoryReservation - The soft limit (in MiB) of memory to reserve for the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_command :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [Text])

The command that is passed to the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_hostname :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Text)

The hostname to use for the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_image :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Text)

The image used to start the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_repositoryCredentials :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe AwsEcsTaskDefinitionContainerDefinitionsRepositoryCredentialsDetails)

The private repository authentication credentials to use.

awsEcsTaskDefinitionContainerDefinitionsDetails_dockerSecurityOptions :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [Text])

A list of strings to provide custom labels for SELinux and AppArmor multi-level security systems.

awsEcsTaskDefinitionContainerDefinitionsDetails_healthCheck :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails)

The container health check command and associated configuration parameters for the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_disableNetworking :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Bool)

Whether to disable networking within the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_volumesFrom :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails])

Data volumes to mount from another container.

awsEcsTaskDefinitionContainerDefinitionsDetails_environment :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsEnvironmentDetails])

The environment variables to pass to a container.

awsEcsTaskDefinitionContainerDefinitionsDetails_secrets :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsSecretsDetails])

The secrets to pass to the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_entryPoint :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [Text])

The entry point that is passed to the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_workingDirectory :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Text)

The working directory in which to run commands inside the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_environmentFiles :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsEnvironmentFilesDetails])

A list of files containing the environment variables to pass to a container.

awsEcsTaskDefinitionContainerDefinitionsDetails_stopTimeout :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Int)

The number of seconds to wait before the container is stopped if it doesn't shut down normally on its own.

awsEcsTaskDefinitionContainerDefinitionsDetails_privileged :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Bool)

Whether the container is given elevated privileges on the host container instance. The elevated privileges are similar to the root user.

awsEcsTaskDefinitionContainerDefinitionsDetails_ulimits :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsUlimitsDetails])

A list of ulimits to set in the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_portMappings :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsPortMappingsDetails])

The list of port mappings for the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_resourceRequirements :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsResourceRequirementsDetails])

The type and amount of a resource to assign to a container. The only supported resource is a GPU.

awsEcsTaskDefinitionContainerDefinitionsDetails_memory :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Int)

The amount (in MiB) of memory to present to the container. If the container attempts to exceed the memory specified here, the container is shut down. The total amount of memory reserved for all containers within a task must be lower than the task memory value, if one is specified.

awsEcsTaskDefinitionContainerDefinitionsDetails_extraHosts :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsExtraHostsDetails])

A list of hostnames and IP address mappings to append to the /etc/hosts file on the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_dockerLabels :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe (HashMap Text Text))

A key-value map of labels to add to the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_user :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Text)

The user to use inside the container.

The value can use one of the following formats.

 user

user : group

 uid

uid : gid

user : gid

uid : group

awsEcsTaskDefinitionContainerDefinitionsDetails_systemControls :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsSystemControlsDetails])

A list of namespaced kernel parameters to set in the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_linuxParameters :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails)

Linux-specific modifications that are applied to the container, such as Linux kernel capabilities.

awsEcsTaskDefinitionContainerDefinitionsDetails_logConfiguration :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails)

The log configuration specification for the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_firelensConfiguration :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe AwsEcsTaskDefinitionContainerDefinitionsFirelensConfigurationDetails)

The FireLens configuration for the container. Specifies and configures a log router for container logs.

awsEcsTaskDefinitionContainerDefinitionsDetails_dnsSearchDomains :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [Text])

A list of DNS search domains that are presented to the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_pseudoTerminal :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Bool)

Whether to allocate a TTY to the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_dependsOn :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsDependsOnDetails])

The dependencies that are defined for container startup and shutdown.

awsEcsTaskDefinitionContainerDefinitionsDetails_dnsServers :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [Text])

A list of DNS servers that are presented to the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_name :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Text)

The name of the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_mountPoints :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsMountPointsDetails])

The mount points for the data volumes in the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_interactive :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Bool)

If set to true, then containerized applications can be deployed that require stdin or a tty to be allocated.

awsEcsTaskDefinitionContainerDefinitionsDetails_startTimeout :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Int)

The number of seconds to wait before giving up on resolving dependencies for a container.

awsEcsTaskDefinitionContainerDefinitionsDetails_readonlyRootFilesystem :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Bool)

Whether the container is given read-only access to its root file system.

awsEcsTaskDefinitionContainerDefinitionsDetails_links :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe [Text])

A list of links for the container in the form container_name:alias . Allows containers to communicate with each other without the need for port mappings.

awsEcsTaskDefinitionContainerDefinitionsDetails_cpu :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Int)

The number of CPU units reserved for the container.

awsEcsTaskDefinitionContainerDefinitionsDetails_essential :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Bool)

Whether the container is essential. All tasks must have at least one essential container.

awsEcsTaskDefinitionContainerDefinitionsDetails_memoryReservation :: Lens' AwsEcsTaskDefinitionContainerDefinitionsDetails (Maybe Int)

The soft limit (in MiB) of memory to reserve for the container.