Amazonka.SecurityHub.Types.AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails

Description

Documentation

data AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails

Linux-specific modifications that are applied to the container, such as

Linux kernel capabilities.

See: newAwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails smart constructor.

Constructors

AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails'

Fields

sharedMemorySize :: Maybe Int

The value for the size (in MiB) of the /dev/shm volume.

tmpfs :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersTmpfsDetails]

The container path, mount options, and size (in MiB) of the tmpfs mount.

initProcessEnabled :: Maybe Bool

Whether to run an init process inside the container that forwards signals and reaps processes.

swappiness :: Maybe Int

Configures the container's memory swappiness behavior. Determines how aggressively pages are swapped. The higher the value, the more aggressive the swappiness. The default is 60.

devices :: Maybe [AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDevicesDetails]

The host devices to expose to the container.

capabilities :: Maybe AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersCapabilitiesDetails

The Linux capabilities for the container that are added to or dropped from the default configuration provided by Docker.

maxSwap :: Maybe Int

The total amount of swap memory (in MiB) that a container can use.

newAwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails :: AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails

Create a value of AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sharedMemorySize:AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails', awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_sharedMemorySize - The value for the size (in MiB) of the /dev/shm volume.

$sel:tmpfs:AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails', awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_tmpfs - The container path, mount options, and size (in MiB) of the tmpfs mount.

$sel:initProcessEnabled:AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails', awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_initProcessEnabled - Whether to run an init process inside the container that forwards signals and reaps processes.

$sel:swappiness:AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails', awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_swappiness - Configures the container's memory swappiness behavior. Determines how aggressively pages are swapped. The higher the value, the more aggressive the swappiness. The default is 60.

$sel:devices:AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails', awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_devices - The host devices to expose to the container.

$sel:capabilities:AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails', awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_capabilities - The Linux capabilities for the container that are added to or dropped from the default configuration provided by Docker.

$sel:maxSwap:AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails', awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_maxSwap - The total amount of swap memory (in MiB) that a container can use.

awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_sharedMemorySize :: Lens' AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails (Maybe Int)

The value for the size (in MiB) of the /dev/shm volume.

awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_tmpfs :: Lens' AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersTmpfsDetails])

The container path, mount options, and size (in MiB) of the tmpfs mount.

awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_initProcessEnabled :: Lens' AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails (Maybe Bool)

Whether to run an init process inside the container that forwards signals and reaps processes.

awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_swappiness :: Lens' AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails (Maybe Int)

Configures the container's memory swappiness behavior. Determines how aggressively pages are swapped. The higher the value, the more aggressive the swappiness. The default is 60.

awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_devices :: Lens' AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails (Maybe [AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDevicesDetails])

The host devices to expose to the container.

awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_capabilities :: Lens' AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails (Maybe AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersCapabilitiesDetails)

The Linux capabilities for the container that are added to or dropped from the default configuration provided by Docker.

awsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails_maxSwap :: Lens' AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails (Maybe Int)

The total amount of swap memory (in MiB) that a container can use.