Amazonka.SecurityHub.Types.AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails

Description

Documentation

data AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails

A data volume to mount from another container.

See: newAwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails smart constructor.

Constructors

AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails'

Fields

sourceContainer :: Maybe Text

The name of another container within the same task definition from which to mount volumes.

readOnly :: Maybe Bool

Whether the container has read-only access to the volume.

newAwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails :: AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails

Create a value of AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceContainer:AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails', awsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails_sourceContainer - The name of another container within the same task definition from which to mount volumes.

$sel:readOnly:AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails', awsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails_readOnly - Whether the container has read-only access to the volume.

awsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails_sourceContainer :: Lens' AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails (Maybe Text)

The name of another container within the same task definition from which to mount volumes.

awsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails_readOnly :: Lens' AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails (Maybe Bool)

Whether the container has read-only access to the volume.