Amazonka.SecurityHub.Types.AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails

Description

Documentation

data AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails

Information about a Docker volume.

See: newAwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails smart constructor.

Constructors

AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails'

Fields

driver :: Maybe Text

The Docker volume driver to use.

driverOpts :: Maybe (HashMap Text Text)

A map of Docker driver-specific options that are passed through.

scope :: Maybe Text

The scope for the Docker volume that determines its lifecycle. Docker volumes that are scoped to a task are provisioned automatically when the task starts and destroyed when the task stops. Docker volumes that are shared persist after the task stops.

labels :: Maybe (HashMap Text Text)

Custom metadata to add to the Docker volume.

autoprovision :: Maybe Bool

Whether to create the Docker volume automatically if it does not already exist.

newAwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails :: AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails

Create a value of AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:driver:AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails', awsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails_driver - The Docker volume driver to use.

$sel:driverOpts:AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails', awsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails_driverOpts - A map of Docker driver-specific options that are passed through.

$sel:scope:AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails', awsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails_scope - The scope for the Docker volume that determines its lifecycle. Docker volumes that are scoped to a task are provisioned automatically when the task starts and destroyed when the task stops. Docker volumes that are shared persist after the task stops.

$sel:labels:AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails', awsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails_labels - Custom metadata to add to the Docker volume.

$sel:autoprovision:AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails', awsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails_autoprovision - Whether to create the Docker volume automatically if it does not already exist.

awsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails_driver :: Lens' AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails (Maybe Text)

The Docker volume driver to use.

awsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails_driverOpts :: Lens' AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails (Maybe (HashMap Text Text))

A map of Docker driver-specific options that are passed through.

awsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails_scope :: Lens' AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails (Maybe Text)

The scope for the Docker volume that determines its lifecycle. Docker volumes that are scoped to a task are provisioned automatically when the task starts and destroyed when the task stops. Docker volumes that are shared persist after the task stops.

awsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails_labels :: Lens' AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails (Maybe (HashMap Text Text))

Custom metadata to add to the Docker volume.

awsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails_autoprovision :: Lens' AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails (Maybe Bool)

Whether to create the Docker volume automatically if it does not already exist.