Amazonka.SecurityHub.Types.AwsEksClusterDetails

Description

Documentation

data AwsEksClusterDetails

Provides details about an Amazon EKS cluster.

See: newAwsEksClusterDetails smart constructor.

Constructors

AwsEksClusterDetails'

Fields

arn :: Maybe Text

The ARN of the cluster.

certificateAuthorityData :: Maybe Text

The certificate authority data for the cluster.

name :: Maybe Text

The name of the cluster.

version :: Maybe Text

The Amazon EKS server version for the cluster.

logging :: Maybe AwsEksClusterLoggingDetails

The logging configuration for the cluster.

endpoint :: Maybe Text

The endpoint for the Amazon EKS API server.

resourcesVpcConfig :: Maybe AwsEksClusterResourcesVpcConfigDetails

The VPC configuration used by the cluster control plane.

clusterStatus :: Maybe Text

The status of the cluster.

roleArn :: Maybe Text

The ARN of the IAM role that provides permissions for the Amazon EKS control plane to make calls to Amazon Web Services API operations on your behalf.

newAwsEksClusterDetails :: AwsEksClusterDetails

Create a value of AwsEksClusterDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:AwsEksClusterDetails', awsEksClusterDetails_arn - The ARN of the cluster.

$sel:certificateAuthorityData:AwsEksClusterDetails', awsEksClusterDetails_certificateAuthorityData - The certificate authority data for the cluster.

$sel:name:AwsEksClusterDetails', awsEksClusterDetails_name - The name of the cluster.

$sel:version:AwsEksClusterDetails', awsEksClusterDetails_version - The Amazon EKS server version for the cluster.

$sel:logging:AwsEksClusterDetails', awsEksClusterDetails_logging - The logging configuration for the cluster.

$sel:endpoint:AwsEksClusterDetails', awsEksClusterDetails_endpoint - The endpoint for the Amazon EKS API server.

$sel:resourcesVpcConfig:AwsEksClusterDetails', awsEksClusterDetails_resourcesVpcConfig - The VPC configuration used by the cluster control plane.

$sel:clusterStatus:AwsEksClusterDetails', awsEksClusterDetails_clusterStatus - The status of the cluster.

$sel:roleArn:AwsEksClusterDetails', awsEksClusterDetails_roleArn - The ARN of the IAM role that provides permissions for the Amazon EKS control plane to make calls to Amazon Web Services API operations on your behalf.

awsEksClusterDetails_arn :: Lens' AwsEksClusterDetails (Maybe Text)

The ARN of the cluster.

awsEksClusterDetails_certificateAuthorityData :: Lens' AwsEksClusterDetails (Maybe Text)

The certificate authority data for the cluster.

awsEksClusterDetails_name :: Lens' AwsEksClusterDetails (Maybe Text)

The name of the cluster.

awsEksClusterDetails_version :: Lens' AwsEksClusterDetails (Maybe Text)

The Amazon EKS server version for the cluster.

awsEksClusterDetails_logging :: Lens' AwsEksClusterDetails (Maybe AwsEksClusterLoggingDetails)

The logging configuration for the cluster.

awsEksClusterDetails_endpoint :: Lens' AwsEksClusterDetails (Maybe Text)

The endpoint for the Amazon EKS API server.

awsEksClusterDetails_resourcesVpcConfig :: Lens' AwsEksClusterDetails (Maybe AwsEksClusterResourcesVpcConfigDetails)

The VPC configuration used by the cluster control plane.

awsEksClusterDetails_clusterStatus :: Lens' AwsEksClusterDetails (Maybe Text)

The status of the cluster.

awsEksClusterDetails_roleArn :: Lens' AwsEksClusterDetails (Maybe Text)

The ARN of the IAM role that provides permissions for the Amazon EKS control plane to make calls to Amazon Web Services API operations on your behalf.