Amazonka.SecurityHub.Types.AwsElasticBeanstalkEnvironmentDetails

Description

Documentation

data AwsElasticBeanstalkEnvironmentDetails

Contains details about an Elastic Beanstalk environment.

See: newAwsElasticBeanstalkEnvironmentDetails smart constructor.

Constructors

AwsElasticBeanstalkEnvironmentDetails'

Fields

status :: Maybe Text

The current operational status of the environment.

cname :: Maybe Text

The URL to the CNAME for this environment.

endpointUrl :: Maybe Text

For load-balanced, autoscaling environments, the URL to the load balancer. For single-instance environments, the IP address of the instance.

optionSettings :: Maybe [AwsElasticBeanstalkEnvironmentOptionSetting]

The configuration setting for the environment.

dateUpdated :: Maybe Text

The date when this environment was last modified.

dateCreated :: Maybe Text

The creation date for this environment.

versionLabel :: Maybe Text

The application version of the environment.

platformArn :: Maybe Text

The ARN of the platform version for the environment.

tier :: Maybe AwsElasticBeanstalkEnvironmentTier

The tier of the environment.

environmentName :: Maybe Text

The name of the environment.

applicationName :: Maybe Text

The name of the application that is associated with the environment.

environmentArn :: Maybe Text

The ARN of the environment.

solutionStackName :: Maybe Text

The name of the solution stack that is deployed with the environment.

environmentId :: Maybe Text

The identifier of the environment.

environmentLinks :: Maybe [AwsElasticBeanstalkEnvironmentEnvironmentLink]

Links to other environments in the same group.

description :: Maybe Text

A description of the environment.

newAwsElasticBeanstalkEnvironmentDetails :: AwsElasticBeanstalkEnvironmentDetails

Create a value of AwsElasticBeanstalkEnvironmentDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_status - The current operational status of the environment.

$sel:cname:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_cname - The URL to the CNAME for this environment.

$sel:endpointUrl:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_endpointUrl - For load-balanced, autoscaling environments, the URL to the load balancer. For single-instance environments, the IP address of the instance.

$sel:optionSettings:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_optionSettings - The configuration setting for the environment.

$sel:dateUpdated:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_dateUpdated - The date when this environment was last modified.

$sel:dateCreated:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_dateCreated - The creation date for this environment.

$sel:versionLabel:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_versionLabel - The application version of the environment.

$sel:platformArn:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_platformArn - The ARN of the platform version for the environment.

$sel:tier:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_tier - The tier of the environment.

$sel:environmentName:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_environmentName - The name of the environment.

$sel:applicationName:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_applicationName - The name of the application that is associated with the environment.

$sel:environmentArn:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_environmentArn - The ARN of the environment.

$sel:solutionStackName:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_solutionStackName - The name of the solution stack that is deployed with the environment.

$sel:environmentId:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_environmentId - The identifier of the environment.

$sel:environmentLinks:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_environmentLinks - Links to other environments in the same group.

$sel:description:AwsElasticBeanstalkEnvironmentDetails', awsElasticBeanstalkEnvironmentDetails_description - A description of the environment.

awsElasticBeanstalkEnvironmentDetails_status :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

The current operational status of the environment.

awsElasticBeanstalkEnvironmentDetails_cname :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

The URL to the CNAME for this environment.

awsElasticBeanstalkEnvironmentDetails_endpointUrl :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

For load-balanced, autoscaling environments, the URL to the load balancer. For single-instance environments, the IP address of the instance.

awsElasticBeanstalkEnvironmentDetails_optionSettings :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe [AwsElasticBeanstalkEnvironmentOptionSetting])

The configuration setting for the environment.

awsElasticBeanstalkEnvironmentDetails_dateUpdated :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

The date when this environment was last modified.

awsElasticBeanstalkEnvironmentDetails_dateCreated :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

The creation date for this environment.

awsElasticBeanstalkEnvironmentDetails_versionLabel :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

The application version of the environment.

awsElasticBeanstalkEnvironmentDetails_platformArn :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

The ARN of the platform version for the environment.

awsElasticBeanstalkEnvironmentDetails_tier :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe AwsElasticBeanstalkEnvironmentTier)

The tier of the environment.

awsElasticBeanstalkEnvironmentDetails_environmentName :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

The name of the environment.

awsElasticBeanstalkEnvironmentDetails_applicationName :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

The name of the application that is associated with the environment.

awsElasticBeanstalkEnvironmentDetails_environmentArn :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

The ARN of the environment.

awsElasticBeanstalkEnvironmentDetails_solutionStackName :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

The name of the solution stack that is deployed with the environment.

awsElasticBeanstalkEnvironmentDetails_environmentId :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

The identifier of the environment.

awsElasticBeanstalkEnvironmentDetails_environmentLinks :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe [AwsElasticBeanstalkEnvironmentEnvironmentLink])

Links to other environments in the same group.

awsElasticBeanstalkEnvironmentDetails_description :: Lens' AwsElasticBeanstalkEnvironmentDetails (Maybe Text)

A description of the environment.