Amazonka.SecurityHub.Types.AwsElasticsearchDomainElasticsearchClusterConfigDetails

Description

Documentation

data AwsElasticsearchDomainElasticsearchClusterConfigDetails

details about the configuration of an OpenSearch cluster.

See: newAwsElasticsearchDomainElasticsearchClusterConfigDetails smart constructor.

Constructors

AwsElasticsearchDomainElasticsearchClusterConfigDetails'

Fields

dedicatedMasterCount :: Maybe Int

The number of instances to use for the master node. If this attribute is specified, then DedicatedMasterEnabled must be true.

dedicatedMasterType :: Maybe Text

The hardware configuration of the computer that hosts the dedicated master node. For example, m3.medium.elasticsearch. If this attribute is specified, then DedicatedMasterEnabled must be true.

dedicatedMasterEnabled :: Maybe Bool

Whether to use a dedicated master node for the Elasticsearch domain. A dedicated master node performs cluster management tasks, but doesn't hold data or respond to data upload requests.

instanceCount :: Maybe Int

The number of data nodes to use in the Elasticsearch domain.

zoneAwarenessEnabled :: Maybe Bool

Whether to enable zone awareness for the Elasticsearch domain. When zone awareness is enabled, OpenSearch allocates the cluster's nodes and replica index shards across Availability Zones in the same Region. This prevents data loss and minimizes downtime if a node or data center fails.

instanceType :: Maybe Text

The instance type for your data nodes. For example, m3.medium.elasticsearch.

zoneAwarenessConfig :: Maybe AwsElasticsearchDomainElasticsearchClusterConfigZoneAwarenessConfigDetails

Configuration options for zone awareness. Provided if ZoneAwarenessEnabled is true.

newAwsElasticsearchDomainElasticsearchClusterConfigDetails :: AwsElasticsearchDomainElasticsearchClusterConfigDetails

Create a value of AwsElasticsearchDomainElasticsearchClusterConfigDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dedicatedMasterCount:AwsElasticsearchDomainElasticsearchClusterConfigDetails', awsElasticsearchDomainElasticsearchClusterConfigDetails_dedicatedMasterCount - The number of instances to use for the master node. If this attribute is specified, then DedicatedMasterEnabled must be true.

$sel:dedicatedMasterType:AwsElasticsearchDomainElasticsearchClusterConfigDetails', awsElasticsearchDomainElasticsearchClusterConfigDetails_dedicatedMasterType - The hardware configuration of the computer that hosts the dedicated master node. For example, m3.medium.elasticsearch. If this attribute is specified, then DedicatedMasterEnabled must be true.

$sel:dedicatedMasterEnabled:AwsElasticsearchDomainElasticsearchClusterConfigDetails', awsElasticsearchDomainElasticsearchClusterConfigDetails_dedicatedMasterEnabled - Whether to use a dedicated master node for the Elasticsearch domain. A dedicated master node performs cluster management tasks, but doesn't hold data or respond to data upload requests.

$sel:instanceCount:AwsElasticsearchDomainElasticsearchClusterConfigDetails', awsElasticsearchDomainElasticsearchClusterConfigDetails_instanceCount - The number of data nodes to use in the Elasticsearch domain.

$sel:zoneAwarenessEnabled:AwsElasticsearchDomainElasticsearchClusterConfigDetails', awsElasticsearchDomainElasticsearchClusterConfigDetails_zoneAwarenessEnabled - Whether to enable zone awareness for the Elasticsearch domain. When zone awareness is enabled, OpenSearch allocates the cluster's nodes and replica index shards across Availability Zones in the same Region. This prevents data loss and minimizes downtime if a node or data center fails.

$sel:instanceType:AwsElasticsearchDomainElasticsearchClusterConfigDetails', awsElasticsearchDomainElasticsearchClusterConfigDetails_instanceType - The instance type for your data nodes. For example, m3.medium.elasticsearch.

$sel:zoneAwarenessConfig:AwsElasticsearchDomainElasticsearchClusterConfigDetails', awsElasticsearchDomainElasticsearchClusterConfigDetails_zoneAwarenessConfig - Configuration options for zone awareness. Provided if ZoneAwarenessEnabled is true.

awsElasticsearchDomainElasticsearchClusterConfigDetails_dedicatedMasterCount :: Lens' AwsElasticsearchDomainElasticsearchClusterConfigDetails (Maybe Int)

The number of instances to use for the master node. If this attribute is specified, then DedicatedMasterEnabled must be true.

awsElasticsearchDomainElasticsearchClusterConfigDetails_dedicatedMasterType :: Lens' AwsElasticsearchDomainElasticsearchClusterConfigDetails (Maybe Text)

The hardware configuration of the computer that hosts the dedicated master node. For example, m3.medium.elasticsearch. If this attribute is specified, then DedicatedMasterEnabled must be true.

awsElasticsearchDomainElasticsearchClusterConfigDetails_dedicatedMasterEnabled :: Lens' AwsElasticsearchDomainElasticsearchClusterConfigDetails (Maybe Bool)

Whether to use a dedicated master node for the Elasticsearch domain. A dedicated master node performs cluster management tasks, but doesn't hold data or respond to data upload requests.

awsElasticsearchDomainElasticsearchClusterConfigDetails_instanceCount :: Lens' AwsElasticsearchDomainElasticsearchClusterConfigDetails (Maybe Int)

The number of data nodes to use in the Elasticsearch domain.

awsElasticsearchDomainElasticsearchClusterConfigDetails_zoneAwarenessEnabled :: Lens' AwsElasticsearchDomainElasticsearchClusterConfigDetails (Maybe Bool)

Whether to enable zone awareness for the Elasticsearch domain. When zone awareness is enabled, OpenSearch allocates the cluster's nodes and replica index shards across Availability Zones in the same Region. This prevents data loss and minimizes downtime if a node or data center fails.

awsElasticsearchDomainElasticsearchClusterConfigDetails_instanceType :: Lens' AwsElasticsearchDomainElasticsearchClusterConfigDetails (Maybe Text)

The instance type for your data nodes. For example, m3.medium.elasticsearch.

awsElasticsearchDomainElasticsearchClusterConfigDetails_zoneAwarenessConfig :: Lens' AwsElasticsearchDomainElasticsearchClusterConfigDetails (Maybe AwsElasticsearchDomainElasticsearchClusterConfigZoneAwarenessConfigDetails)

Configuration options for zone awareness. Provided if ZoneAwarenessEnabled is true.