Amazonka.SecurityHub.Types.AwsElasticsearchDomainServiceSoftwareOptions

Description

Documentation

data AwsElasticsearchDomainServiceSoftwareOptions

Information about the state of the domain relative to the latest service software.

See: newAwsElasticsearchDomainServiceSoftwareOptions smart constructor.

Constructors

AwsElasticsearchDomainServiceSoftwareOptions'

Fields

automatedUpdateDate :: Maybe Text

The epoch time when the deployment window closes for required updates. After this time, Amazon OpenSearch Service schedules the software upgrade automatically.

currentVersion :: Maybe Text

The version of the service software that is currently installed on the domain.

updateStatus :: Maybe Text

The status of the service software update.

cancellable :: Maybe Bool

Whether a request to update the domain can be canceled.

updateAvailable :: Maybe Bool

Whether a service software update is available for the domain.

description :: Maybe Text

A more detailed description of the service software status.

newVersion' :: Maybe Text

The most recent version of the service software.

newAwsElasticsearchDomainServiceSoftwareOptions :: AwsElasticsearchDomainServiceSoftwareOptions

Create a value of AwsElasticsearchDomainServiceSoftwareOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:automatedUpdateDate:AwsElasticsearchDomainServiceSoftwareOptions', awsElasticsearchDomainServiceSoftwareOptions_automatedUpdateDate - The epoch time when the deployment window closes for required updates. After this time, Amazon OpenSearch Service schedules the software upgrade automatically.

$sel:currentVersion:AwsElasticsearchDomainServiceSoftwareOptions', awsElasticsearchDomainServiceSoftwareOptions_currentVersion - The version of the service software that is currently installed on the domain.

$sel:updateStatus:AwsElasticsearchDomainServiceSoftwareOptions', awsElasticsearchDomainServiceSoftwareOptions_updateStatus - The status of the service software update.

$sel:cancellable:AwsElasticsearchDomainServiceSoftwareOptions', awsElasticsearchDomainServiceSoftwareOptions_cancellable - Whether a request to update the domain can be canceled.

$sel:updateAvailable:AwsElasticsearchDomainServiceSoftwareOptions', awsElasticsearchDomainServiceSoftwareOptions_updateAvailable - Whether a service software update is available for the domain.

$sel:description:AwsElasticsearchDomainServiceSoftwareOptions', awsElasticsearchDomainServiceSoftwareOptions_description - A more detailed description of the service software status.

$sel:newVersion':AwsElasticsearchDomainServiceSoftwareOptions', awsElasticsearchDomainServiceSoftwareOptions_newVersion - The most recent version of the service software.

awsElasticsearchDomainServiceSoftwareOptions_automatedUpdateDate :: Lens' AwsElasticsearchDomainServiceSoftwareOptions (Maybe Text)

The epoch time when the deployment window closes for required updates. After this time, Amazon OpenSearch Service schedules the software upgrade automatically.

awsElasticsearchDomainServiceSoftwareOptions_currentVersion :: Lens' AwsElasticsearchDomainServiceSoftwareOptions (Maybe Text)

The version of the service software that is currently installed on the domain.

awsElasticsearchDomainServiceSoftwareOptions_updateStatus :: Lens' AwsElasticsearchDomainServiceSoftwareOptions (Maybe Text)

The status of the service software update.

awsElasticsearchDomainServiceSoftwareOptions_cancellable :: Lens' AwsElasticsearchDomainServiceSoftwareOptions (Maybe Bool)

Whether a request to update the domain can be canceled.

awsElasticsearchDomainServiceSoftwareOptions_updateAvailable :: Lens' AwsElasticsearchDomainServiceSoftwareOptions (Maybe Bool)

Whether a service software update is available for the domain.

awsElasticsearchDomainServiceSoftwareOptions_description :: Lens' AwsElasticsearchDomainServiceSoftwareOptions (Maybe Text)

A more detailed description of the service software status.

awsElasticsearchDomainServiceSoftwareOptions_newVersion :: Lens' AwsElasticsearchDomainServiceSoftwareOptions (Maybe Text)

The most recent version of the service software.