Amazonka.SecurityHub.Types.AwsElbAppCookieStickinessPolicy

Description

Documentation

data AwsElbAppCookieStickinessPolicy

Contains information about a stickiness policy that was created using CreateAppCookieStickinessPolicy.

See: newAwsElbAppCookieStickinessPolicy smart constructor.

Constructors

AwsElbAppCookieStickinessPolicy'

Fields

policyName :: Maybe Text

The mnemonic name for the policy being created. The name must be unique within the set of policies for the load balancer.

cookieName :: Maybe Text

The name of the application cookie used for stickiness.

newAwsElbAppCookieStickinessPolicy :: AwsElbAppCookieStickinessPolicy

Create a value of AwsElbAppCookieStickinessPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyName:AwsElbAppCookieStickinessPolicy', awsElbAppCookieStickinessPolicy_policyName - The mnemonic name for the policy being created. The name must be unique within the set of policies for the load balancer.

$sel:cookieName:AwsElbAppCookieStickinessPolicy', awsElbAppCookieStickinessPolicy_cookieName - The name of the application cookie used for stickiness.

awsElbAppCookieStickinessPolicy_policyName :: Lens' AwsElbAppCookieStickinessPolicy (Maybe Text)

The mnemonic name for the policy being created. The name must be unique within the set of policies for the load balancer.

awsElbAppCookieStickinessPolicy_cookieName :: Lens' AwsElbAppCookieStickinessPolicy (Maybe Text)

The name of the application cookie used for stickiness.