Amazonka.SecurityHub.Types.AwsElbLoadBalancerDetails

Description

Documentation

data AwsElbLoadBalancerDetails

Contains details about a Classic Load Balancer.

See: newAwsElbLoadBalancerDetails smart constructor.

Constructors

AwsElbLoadBalancerDetails'

Fields

sourceSecurityGroup :: Maybe AwsElbLoadBalancerSourceSecurityGroup

Information about the security group for the load balancer. This is the security group that is used for inbound rules.

canonicalHostedZoneName :: Maybe Text

The name of the Amazon Route 53 hosted zone for the load balancer.

securityGroups :: Maybe [Text]

The security groups for the load balancer. Only provided if the load balancer is in a VPC.

healthCheck :: Maybe AwsElbLoadBalancerHealthCheck

Information about the health checks that are conducted on the load balancer.

loadBalancerName :: Maybe Text

The name of the load balancer.

loadBalancerAttributes :: Maybe AwsElbLoadBalancerAttributes

The attributes for a load balancer.

createdTime :: Maybe Text

Indicates when the load balancer was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

vpcId :: Maybe Text

The identifier of the VPC for the load balancer.

subnets :: Maybe [Text]

The list of subnet identifiers for the load balancer.

availabilityZones :: Maybe [Text]

The list of Availability Zones for the load balancer.

backendServerDescriptions :: Maybe [AwsElbLoadBalancerBackendServerDescription]

Information about the configuration of the EC2 instances.

canonicalHostedZoneNameID :: Maybe Text

The ID of the Amazon Route 53 hosted zone for the load balancer.

instances :: Maybe [AwsElbLoadBalancerInstance]

List of EC2 instances for the load balancer.

scheme :: Maybe Text

The type of load balancer. Only provided if the load balancer is in a VPC.

If Scheme is internet-facing, the load balancer has a public DNS name that resolves to a public IP address.

If Scheme is internal, the load balancer has a public DNS name that resolves to a private IP address.

listenerDescriptions :: Maybe [AwsElbLoadBalancerListenerDescription]

The policies that are enabled for the load balancer listeners.

dnsName :: Maybe Text

The DNS name of the load balancer.

policies :: Maybe AwsElbLoadBalancerPolicies

The policies for a load balancer.

newAwsElbLoadBalancerDetails :: AwsElbLoadBalancerDetails

Create a value of AwsElbLoadBalancerDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceSecurityGroup:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_sourceSecurityGroup - Information about the security group for the load balancer. This is the security group that is used for inbound rules.

$sel:canonicalHostedZoneName:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_canonicalHostedZoneName - The name of the Amazon Route 53 hosted zone for the load balancer.

$sel:securityGroups:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_securityGroups - The security groups for the load balancer. Only provided if the load balancer is in a VPC.

$sel:healthCheck:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_healthCheck - Information about the health checks that are conducted on the load balancer.

$sel:loadBalancerName:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_loadBalancerName - The name of the load balancer.

$sel:loadBalancerAttributes:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_loadBalancerAttributes - The attributes for a load balancer.

$sel:createdTime:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_createdTime - Indicates when the load balancer was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:vpcId:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_vpcId - The identifier of the VPC for the load balancer.

$sel:subnets:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_subnets - The list of subnet identifiers for the load balancer.

$sel:availabilityZones:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_availabilityZones - The list of Availability Zones for the load balancer.

$sel:backendServerDescriptions:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_backendServerDescriptions - Information about the configuration of the EC2 instances.

$sel:canonicalHostedZoneNameID:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_canonicalHostedZoneNameID - The ID of the Amazon Route 53 hosted zone for the load balancer.

$sel:instances:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_instances - List of EC2 instances for the load balancer.

$sel:scheme:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_scheme - The type of load balancer. Only provided if the load balancer is in a VPC.

If Scheme is internet-facing, the load balancer has a public DNS name that resolves to a public IP address.

If Scheme is internal, the load balancer has a public DNS name that resolves to a private IP address.

$sel:listenerDescriptions:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_listenerDescriptions - The policies that are enabled for the load balancer listeners.

$sel:dnsName:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_dnsName - The DNS name of the load balancer.

$sel:policies:AwsElbLoadBalancerDetails', awsElbLoadBalancerDetails_policies - The policies for a load balancer.

awsElbLoadBalancerDetails_sourceSecurityGroup :: Lens' AwsElbLoadBalancerDetails (Maybe AwsElbLoadBalancerSourceSecurityGroup)

Information about the security group for the load balancer. This is the security group that is used for inbound rules.

awsElbLoadBalancerDetails_canonicalHostedZoneName :: Lens' AwsElbLoadBalancerDetails (Maybe Text)

The name of the Amazon Route 53 hosted zone for the load balancer.

awsElbLoadBalancerDetails_securityGroups :: Lens' AwsElbLoadBalancerDetails (Maybe [Text])

The security groups for the load balancer. Only provided if the load balancer is in a VPC.

awsElbLoadBalancerDetails_healthCheck :: Lens' AwsElbLoadBalancerDetails (Maybe AwsElbLoadBalancerHealthCheck)

Information about the health checks that are conducted on the load balancer.

awsElbLoadBalancerDetails_loadBalancerName :: Lens' AwsElbLoadBalancerDetails (Maybe Text)

The name of the load balancer.

awsElbLoadBalancerDetails_loadBalancerAttributes :: Lens' AwsElbLoadBalancerDetails (Maybe AwsElbLoadBalancerAttributes)

The attributes for a load balancer.

awsElbLoadBalancerDetails_createdTime :: Lens' AwsElbLoadBalancerDetails (Maybe Text)

Indicates when the load balancer was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsElbLoadBalancerDetails_vpcId :: Lens' AwsElbLoadBalancerDetails (Maybe Text)

The identifier of the VPC for the load balancer.

awsElbLoadBalancerDetails_subnets :: Lens' AwsElbLoadBalancerDetails (Maybe [Text])

The list of subnet identifiers for the load balancer.

awsElbLoadBalancerDetails_availabilityZones :: Lens' AwsElbLoadBalancerDetails (Maybe [Text])

The list of Availability Zones for the load balancer.

awsElbLoadBalancerDetails_backendServerDescriptions :: Lens' AwsElbLoadBalancerDetails (Maybe [AwsElbLoadBalancerBackendServerDescription])

Information about the configuration of the EC2 instances.

awsElbLoadBalancerDetails_canonicalHostedZoneNameID :: Lens' AwsElbLoadBalancerDetails (Maybe Text)

The ID of the Amazon Route 53 hosted zone for the load balancer.

awsElbLoadBalancerDetails_instances :: Lens' AwsElbLoadBalancerDetails (Maybe [AwsElbLoadBalancerInstance])

List of EC2 instances for the load balancer.

awsElbLoadBalancerDetails_scheme :: Lens' AwsElbLoadBalancerDetails (Maybe Text)

The type of load balancer. Only provided if the load balancer is in a VPC.

If Scheme is internet-facing, the load balancer has a public DNS name that resolves to a public IP address.

If Scheme is internal, the load balancer has a public DNS name that resolves to a private IP address.

awsElbLoadBalancerDetails_listenerDescriptions :: Lens' AwsElbLoadBalancerDetails (Maybe [AwsElbLoadBalancerListenerDescription])

The policies that are enabled for the load balancer listeners.

awsElbLoadBalancerDetails_dnsName :: Lens' AwsElbLoadBalancerDetails (Maybe Text)

The DNS name of the load balancer.

awsElbLoadBalancerDetails_policies :: Lens' AwsElbLoadBalancerDetails (Maybe AwsElbLoadBalancerPolicies)

The policies for a load balancer.