Amazonka.SecurityHub.Types.AwsElbLoadBalancerHealthCheck

Description

Documentation

data AwsElbLoadBalancerHealthCheck

Contains information about the health checks that are conducted on the load balancer.

See: newAwsElbLoadBalancerHealthCheck smart constructor.

Constructors

AwsElbLoadBalancerHealthCheck'

Fields

healthyThreshold :: Maybe Int

The number of consecutive health check successes required before the instance is moved to the Healthy state.

interval :: Maybe Int

The approximate interval, in seconds, between health checks of an individual instance.

timeout :: Maybe Int

The amount of time, in seconds, during which no response means a failed health check.

unhealthyThreshold :: Maybe Int

The number of consecutive health check failures that must occur before the instance is moved to the Unhealthy state.

target :: Maybe Text

The instance that is being checked. The target specifies the protocol and port. The available protocols are TCP, SSL, HTTP, and HTTPS. The range of valid ports is 1 through 65535.

For the HTTP and HTTPS protocols, the target also specifies the ping path.

For the TCP protocol, the target is specified as TCP: <port> .

For the SSL protocol, the target is specified as SSL.<port> .

For the HTTP and HTTPS protocols, the target is specified as <protocol>:<port>/<path to ping> .

newAwsElbLoadBalancerHealthCheck :: AwsElbLoadBalancerHealthCheck

Create a value of AwsElbLoadBalancerHealthCheck with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:healthyThreshold:AwsElbLoadBalancerHealthCheck', awsElbLoadBalancerHealthCheck_healthyThreshold - The number of consecutive health check successes required before the instance is moved to the Healthy state.

$sel:interval:AwsElbLoadBalancerHealthCheck', awsElbLoadBalancerHealthCheck_interval - The approximate interval, in seconds, between health checks of an individual instance.

$sel:timeout:AwsElbLoadBalancerHealthCheck', awsElbLoadBalancerHealthCheck_timeout - The amount of time, in seconds, during which no response means a failed health check.

$sel:unhealthyThreshold:AwsElbLoadBalancerHealthCheck', awsElbLoadBalancerHealthCheck_unhealthyThreshold - The number of consecutive health check failures that must occur before the instance is moved to the Unhealthy state.

$sel:target:AwsElbLoadBalancerHealthCheck', awsElbLoadBalancerHealthCheck_target - The instance that is being checked. The target specifies the protocol and port. The available protocols are TCP, SSL, HTTP, and HTTPS. The range of valid ports is 1 through 65535.

For the HTTP and HTTPS protocols, the target also specifies the ping path.

For the TCP protocol, the target is specified as TCP: <port> .

For the SSL protocol, the target is specified as SSL.<port> .

For the HTTP and HTTPS protocols, the target is specified as <protocol>:<port>/<path to ping> .

awsElbLoadBalancerHealthCheck_healthyThreshold :: Lens' AwsElbLoadBalancerHealthCheck (Maybe Int)

The number of consecutive health check successes required before the instance is moved to the Healthy state.

awsElbLoadBalancerHealthCheck_interval :: Lens' AwsElbLoadBalancerHealthCheck (Maybe Int)

The approximate interval, in seconds, between health checks of an individual instance.

awsElbLoadBalancerHealthCheck_timeout :: Lens' AwsElbLoadBalancerHealthCheck (Maybe Int)

The amount of time, in seconds, during which no response means a failed health check.

awsElbLoadBalancerHealthCheck_unhealthyThreshold :: Lens' AwsElbLoadBalancerHealthCheck (Maybe Int)

The number of consecutive health check failures that must occur before the instance is moved to the Unhealthy state.

awsElbLoadBalancerHealthCheck_target :: Lens' AwsElbLoadBalancerHealthCheck (Maybe Text)

The instance that is being checked. The target specifies the protocol and port. The available protocols are TCP, SSL, HTTP, and HTTPS. The range of valid ports is 1 through 65535.

For the HTTP and HTTPS protocols, the target also specifies the ping path.

For the TCP protocol, the target is specified as TCP: <port> .

For the SSL protocol, the target is specified as SSL.<port> .

For the HTTP and HTTPS protocols, the target is specified as <protocol>:<port>/<path to ping> .