Amazonka.SecurityHub.Types.AwsIamAccessKeyDetails

Description

Documentation

data AwsIamAccessKeyDetails

IAM access key details related to a finding.

See: newAwsIamAccessKeyDetails smart constructor.

Constructors

AwsIamAccessKeyDetails'

Fields

status :: Maybe AwsIamAccessKeyStatus

The status of the IAM access key related to a finding.

principalId :: Maybe Text

The ID of the principal associated with an access key.

principalType :: Maybe Text

The type of principal associated with an access key.

principalName :: Maybe Text

The name of the principal.

createdAt :: Maybe Text

Indicates when the IAM access key was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

userName :: Maybe Text

The user associated with the IAM access key related to a finding.

The UserName parameter has been replaced with the PrincipalName parameter because access keys can also be assigned to principals that are not IAM users.

sessionContext :: Maybe AwsIamAccessKeySessionContext

Information about the session that the key was used for.

accountId :: Maybe Text

The Amazon Web Services account ID of the account for the key.

accessKeyId :: Maybe Text

The identifier of the access key.

newAwsIamAccessKeyDetails :: AwsIamAccessKeyDetails

Create a value of AwsIamAccessKeyDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AwsIamAccessKeyDetails', awsIamAccessKeyDetails_status - The status of the IAM access key related to a finding.

$sel:principalId:AwsIamAccessKeyDetails', awsIamAccessKeyDetails_principalId - The ID of the principal associated with an access key.

$sel:principalType:AwsIamAccessKeyDetails', awsIamAccessKeyDetails_principalType - The type of principal associated with an access key.

$sel:principalName:AwsIamAccessKeyDetails', awsIamAccessKeyDetails_principalName - The name of the principal.

$sel:createdAt:AwsIamAccessKeyDetails', awsIamAccessKeyDetails_createdAt - Indicates when the IAM access key was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:userName:AwsIamAccessKeyDetails', awsIamAccessKeyDetails_userName - The user associated with the IAM access key related to a finding.

The UserName parameter has been replaced with the PrincipalName parameter because access keys can also be assigned to principals that are not IAM users.

$sel:sessionContext:AwsIamAccessKeyDetails', awsIamAccessKeyDetails_sessionContext - Information about the session that the key was used for.

$sel:accountId:AwsIamAccessKeyDetails', awsIamAccessKeyDetails_accountId - The Amazon Web Services account ID of the account for the key.

$sel:accessKeyId:AwsIamAccessKeyDetails', awsIamAccessKeyDetails_accessKeyId - The identifier of the access key.

awsIamAccessKeyDetails_status :: Lens' AwsIamAccessKeyDetails (Maybe AwsIamAccessKeyStatus)

The status of the IAM access key related to a finding.

awsIamAccessKeyDetails_principalId :: Lens' AwsIamAccessKeyDetails (Maybe Text)

The ID of the principal associated with an access key.

awsIamAccessKeyDetails_principalType :: Lens' AwsIamAccessKeyDetails (Maybe Text)

The type of principal associated with an access key.

awsIamAccessKeyDetails_principalName :: Lens' AwsIamAccessKeyDetails (Maybe Text)

The name of the principal.

awsIamAccessKeyDetails_createdAt :: Lens' AwsIamAccessKeyDetails (Maybe Text)

Indicates when the IAM access key was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsIamAccessKeyDetails_userName :: Lens' AwsIamAccessKeyDetails (Maybe Text)

The user associated with the IAM access key related to a finding.

The UserName parameter has been replaced with the PrincipalName parameter because access keys can also be assigned to principals that are not IAM users.

awsIamAccessKeyDetails_sessionContext :: Lens' AwsIamAccessKeyDetails (Maybe AwsIamAccessKeySessionContext)

Information about the session that the key was used for.

awsIamAccessKeyDetails_accountId :: Lens' AwsIamAccessKeyDetails (Maybe Text)

The Amazon Web Services account ID of the account for the key.

awsIamAccessKeyDetails_accessKeyId :: Lens' AwsIamAccessKeyDetails (Maybe Text)

The identifier of the access key.