Amazonka.SecurityHub.Types.AwsIamAccessKeySessionContextSessionIssuer

Description

Documentation

data AwsIamAccessKeySessionContextSessionIssuer

Information about the entity that created the session.

See: newAwsIamAccessKeySessionContextSessionIssuer smart constructor.

Constructors

AwsIamAccessKeySessionContextSessionIssuer'

Fields

principalId :: Maybe Text

The principal ID of the principal (user, role, or group) that created the session.

arn :: Maybe Text

The ARN of the session.

userName :: Maybe Text

The name of the principal that created the session.

accountId :: Maybe Text

The identifier of the Amazon Web Services account that created the session.

type' :: Maybe Text

The type of principal (user, role, or group) that created the session.

newAwsIamAccessKeySessionContextSessionIssuer :: AwsIamAccessKeySessionContextSessionIssuer

Create a value of AwsIamAccessKeySessionContextSessionIssuer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principalId:AwsIamAccessKeySessionContextSessionIssuer', awsIamAccessKeySessionContextSessionIssuer_principalId - The principal ID of the principal (user, role, or group) that created the session.

$sel:arn:AwsIamAccessKeySessionContextSessionIssuer', awsIamAccessKeySessionContextSessionIssuer_arn - The ARN of the session.

$sel:userName:AwsIamAccessKeySessionContextSessionIssuer', awsIamAccessKeySessionContextSessionIssuer_userName - The name of the principal that created the session.

$sel:accountId:AwsIamAccessKeySessionContextSessionIssuer', awsIamAccessKeySessionContextSessionIssuer_accountId - The identifier of the Amazon Web Services account that created the session.

$sel:type':AwsIamAccessKeySessionContextSessionIssuer', awsIamAccessKeySessionContextSessionIssuer_type - The type of principal (user, role, or group) that created the session.

awsIamAccessKeySessionContextSessionIssuer_principalId :: Lens' AwsIamAccessKeySessionContextSessionIssuer (Maybe Text)

The principal ID of the principal (user, role, or group) that created the session.

awsIamAccessKeySessionContextSessionIssuer_arn :: Lens' AwsIamAccessKeySessionContextSessionIssuer (Maybe Text)

The ARN of the session.

awsIamAccessKeySessionContextSessionIssuer_userName :: Lens' AwsIamAccessKeySessionContextSessionIssuer (Maybe Text)

The name of the principal that created the session.

awsIamAccessKeySessionContextSessionIssuer_accountId :: Lens' AwsIamAccessKeySessionContextSessionIssuer (Maybe Text)

The identifier of the Amazon Web Services account that created the session.

awsIamAccessKeySessionContextSessionIssuer_type :: Lens' AwsIamAccessKeySessionContextSessionIssuer (Maybe Text)

The type of principal (user, role, or group) that created the session.