Amazonka.SecurityHub.Types.AwsIamPolicyDetails

Description

Documentation

data AwsIamPolicyDetails

Represents an IAM permissions policy.

See: newAwsIamPolicyDetails smart constructor.

Constructors

AwsIamPolicyDetails'

Fields

policyName :: Maybe Text

The name of the policy.

updateDate :: Maybe Text

When the policy was most recently updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

policyId :: Maybe Text

The unique identifier of the policy.

path :: Maybe Text

The path to the policy.

policyVersionList :: Maybe [AwsIamPolicyVersion]

List of versions of the policy.

createDate :: Maybe Text

When the policy was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

isAttachable :: Maybe Bool

Whether the policy can be attached to a user, group, or role.

permissionsBoundaryUsageCount :: Maybe Int

The number of users and roles that use the policy to set the permissions boundary.

defaultVersionId :: Maybe Text

The identifier of the default version of the policy.

attachmentCount :: Maybe Int

The number of users, groups, and roles that the policy is attached to.

description :: Maybe Text

A description of the policy.

newAwsIamPolicyDetails :: AwsIamPolicyDetails

Create a value of AwsIamPolicyDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyName:AwsIamPolicyDetails', awsIamPolicyDetails_policyName - The name of the policy.

$sel:updateDate:AwsIamPolicyDetails', awsIamPolicyDetails_updateDate - When the policy was most recently updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:policyId:AwsIamPolicyDetails', awsIamPolicyDetails_policyId - The unique identifier of the policy.

$sel:path:AwsIamPolicyDetails', awsIamPolicyDetails_path - The path to the policy.

$sel:policyVersionList:AwsIamPolicyDetails', awsIamPolicyDetails_policyVersionList - List of versions of the policy.

$sel:createDate:AwsIamPolicyDetails', awsIamPolicyDetails_createDate - When the policy was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:isAttachable:AwsIamPolicyDetails', awsIamPolicyDetails_isAttachable - Whether the policy can be attached to a user, group, or role.

$sel:permissionsBoundaryUsageCount:AwsIamPolicyDetails', awsIamPolicyDetails_permissionsBoundaryUsageCount - The number of users and roles that use the policy to set the permissions boundary.

$sel:defaultVersionId:AwsIamPolicyDetails', awsIamPolicyDetails_defaultVersionId - The identifier of the default version of the policy.

$sel:attachmentCount:AwsIamPolicyDetails', awsIamPolicyDetails_attachmentCount - The number of users, groups, and roles that the policy is attached to.

$sel:description:AwsIamPolicyDetails', awsIamPolicyDetails_description - A description of the policy.

awsIamPolicyDetails_policyName :: Lens' AwsIamPolicyDetails (Maybe Text)

The name of the policy.

awsIamPolicyDetails_updateDate :: Lens' AwsIamPolicyDetails (Maybe Text)

When the policy was most recently updated.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsIamPolicyDetails_policyId :: Lens' AwsIamPolicyDetails (Maybe Text)

The unique identifier of the policy.

awsIamPolicyDetails_path :: Lens' AwsIamPolicyDetails (Maybe Text)

The path to the policy.

awsIamPolicyDetails_policyVersionList :: Lens' AwsIamPolicyDetails (Maybe [AwsIamPolicyVersion])

List of versions of the policy.

awsIamPolicyDetails_createDate :: Lens' AwsIamPolicyDetails (Maybe Text)

When the policy was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsIamPolicyDetails_isAttachable :: Lens' AwsIamPolicyDetails (Maybe Bool)

Whether the policy can be attached to a user, group, or role.

awsIamPolicyDetails_permissionsBoundaryUsageCount :: Lens' AwsIamPolicyDetails (Maybe Int)

The number of users and roles that use the policy to set the permissions boundary.

awsIamPolicyDetails_defaultVersionId :: Lens' AwsIamPolicyDetails (Maybe Text)

The identifier of the default version of the policy.

awsIamPolicyDetails_attachmentCount :: Lens' AwsIamPolicyDetails (Maybe Int)

The number of users, groups, and roles that the policy is attached to.

awsIamPolicyDetails_description :: Lens' AwsIamPolicyDetails (Maybe Text)

A description of the policy.