Amazonka.SecurityHub.Types.AwsIamRoleDetails

Description

Documentation

data AwsIamRoleDetails

Contains information about an IAM role, including all of the role's policies.

See: newAwsIamRoleDetails smart constructor.

Constructors

AwsIamRoleDetails'

Fields

maxSessionDuration :: Maybe Int

The maximum session duration (in seconds) that you want to set for the specified role.

assumeRolePolicyDocument :: Maybe Text

The trust policy that grants permission to assume the role.

path :: Maybe Text

The path to the role.

instanceProfileList :: Maybe [AwsIamInstanceProfile]

The list of instance profiles that contain this role.

createDate :: Maybe Text

Indicates when the role was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

roleName :: Maybe Text

The friendly name that identifies the role.

roleId :: Maybe Text

The stable and unique string identifying the role.

permissionsBoundary :: Maybe AwsIamPermissionsBoundary

rolePolicyList :: Maybe [AwsIamRolePolicy]

The list of inline policies that are embedded in the role.

attachedManagedPolicies :: Maybe [AwsIamAttachedManagedPolicy]

The list of the managed policies that are attached to the role.

newAwsIamRoleDetails :: AwsIamRoleDetails

Create a value of AwsIamRoleDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxSessionDuration:AwsIamRoleDetails', awsIamRoleDetails_maxSessionDuration - The maximum session duration (in seconds) that you want to set for the specified role.

$sel:assumeRolePolicyDocument:AwsIamRoleDetails', awsIamRoleDetails_assumeRolePolicyDocument - The trust policy that grants permission to assume the role.

$sel:path:AwsIamRoleDetails', awsIamRoleDetails_path - The path to the role.

$sel:instanceProfileList:AwsIamRoleDetails', awsIamRoleDetails_instanceProfileList - The list of instance profiles that contain this role.

$sel:createDate:AwsIamRoleDetails', awsIamRoleDetails_createDate - Indicates when the role was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:roleName:AwsIamRoleDetails', awsIamRoleDetails_roleName - The friendly name that identifies the role.

$sel:roleId:AwsIamRoleDetails', awsIamRoleDetails_roleId - The stable and unique string identifying the role.

$sel:permissionsBoundary:AwsIamRoleDetails', awsIamRoleDetails_permissionsBoundary - Undocumented member.

$sel:rolePolicyList:AwsIamRoleDetails', awsIamRoleDetails_rolePolicyList - The list of inline policies that are embedded in the role.

$sel:attachedManagedPolicies:AwsIamRoleDetails', awsIamRoleDetails_attachedManagedPolicies - The list of the managed policies that are attached to the role.

awsIamRoleDetails_maxSessionDuration :: Lens' AwsIamRoleDetails (Maybe Int)

The maximum session duration (in seconds) that you want to set for the specified role.

awsIamRoleDetails_assumeRolePolicyDocument :: Lens' AwsIamRoleDetails (Maybe Text)

The trust policy that grants permission to assume the role.

awsIamRoleDetails_path :: Lens' AwsIamRoleDetails (Maybe Text)

The path to the role.

awsIamRoleDetails_instanceProfileList :: Lens' AwsIamRoleDetails (Maybe [AwsIamInstanceProfile])

The list of instance profiles that contain this role.

awsIamRoleDetails_createDate :: Lens' AwsIamRoleDetails (Maybe Text)

Indicates when the role was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsIamRoleDetails_roleName :: Lens' AwsIamRoleDetails (Maybe Text)

The friendly name that identifies the role.

awsIamRoleDetails_roleId :: Lens' AwsIamRoleDetails (Maybe Text)

The stable and unique string identifying the role.

awsIamRoleDetails_permissionsBoundary :: Lens' AwsIamRoleDetails (Maybe AwsIamPermissionsBoundary)

Undocumented member.

awsIamRoleDetails_rolePolicyList :: Lens' AwsIamRoleDetails (Maybe [AwsIamRolePolicy])

The list of inline policies that are embedded in the role.

awsIamRoleDetails_attachedManagedPolicies :: Lens' AwsIamRoleDetails (Maybe [AwsIamAttachedManagedPolicy])

The list of the managed policies that are attached to the role.