Amazonka.SecurityHub.Types.AwsKmsKeyDetails

Description

Documentation

data AwsKmsKeyDetails

Contains metadata about an KMS key.

See: newAwsKmsKeyDetails smart constructor.

Constructors

AwsKmsKeyDetails'

Fields

origin :: Maybe Text

The source of the KMS key material.

When this value is AWS_KMS, KMS created the key material.

When this value is EXTERNAL, the key material was imported from your existing key management infrastructure or the KMS key lacks key material.

When this value is AWS_CLOUDHSM, the key material was created in the CloudHSM cluster associated with a custom key store.

keyManager :: Maybe Text

The manager of the KMS key. KMS keys in your Amazon Web Services account are either customer managed or Amazon Web Services managed.

keyId :: Maybe Text

The globally unique identifier for the KMS key.

keyState :: Maybe Text

The state of the KMS key.

aWSAccountId :: Maybe Text

The twelve-digit account ID of the Amazon Web Services account that owns the KMS key.

keyRotationStatus :: Maybe Bool

Whether the key has key rotation enabled.

creationDate :: Maybe Double

Indicates when the KMS key was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

description :: Maybe Text

A description of the key.

newAwsKmsKeyDetails :: AwsKmsKeyDetails

Create a value of AwsKmsKeyDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:origin:AwsKmsKeyDetails', awsKmsKeyDetails_origin - The source of the KMS key material.

When this value is AWS_KMS, KMS created the key material.

When this value is EXTERNAL, the key material was imported from your existing key management infrastructure or the KMS key lacks key material.

When this value is AWS_CLOUDHSM, the key material was created in the CloudHSM cluster associated with a custom key store.

$sel:keyManager:AwsKmsKeyDetails', awsKmsKeyDetails_keyManager - The manager of the KMS key. KMS keys in your Amazon Web Services account are either customer managed or Amazon Web Services managed.

$sel:keyId:AwsKmsKeyDetails', awsKmsKeyDetails_keyId - The globally unique identifier for the KMS key.

$sel:keyState:AwsKmsKeyDetails', awsKmsKeyDetails_keyState - The state of the KMS key.

$sel:aWSAccountId:AwsKmsKeyDetails', awsKmsKeyDetails_aWSAccountId - The twelve-digit account ID of the Amazon Web Services account that owns the KMS key.

$sel:keyRotationStatus:AwsKmsKeyDetails', awsKmsKeyDetails_keyRotationStatus - Whether the key has key rotation enabled.

$sel:creationDate:AwsKmsKeyDetails', awsKmsKeyDetails_creationDate - Indicates when the KMS key was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:description:AwsKmsKeyDetails', awsKmsKeyDetails_description - A description of the key.

awsKmsKeyDetails_origin :: Lens' AwsKmsKeyDetails (Maybe Text)

The source of the KMS key material.

When this value is AWS_KMS, KMS created the key material.

When this value is EXTERNAL, the key material was imported from your existing key management infrastructure or the KMS key lacks key material.

When this value is AWS_CLOUDHSM, the key material was created in the CloudHSM cluster associated with a custom key store.

awsKmsKeyDetails_keyManager :: Lens' AwsKmsKeyDetails (Maybe Text)

The manager of the KMS key. KMS keys in your Amazon Web Services account are either customer managed or Amazon Web Services managed.

awsKmsKeyDetails_keyId :: Lens' AwsKmsKeyDetails (Maybe Text)

The globally unique identifier for the KMS key.

awsKmsKeyDetails_keyState :: Lens' AwsKmsKeyDetails (Maybe Text)

The state of the KMS key.

awsKmsKeyDetails_aWSAccountId :: Lens' AwsKmsKeyDetails (Maybe Text)

The twelve-digit account ID of the Amazon Web Services account that owns the KMS key.

awsKmsKeyDetails_keyRotationStatus :: Lens' AwsKmsKeyDetails (Maybe Bool)

Whether the key has key rotation enabled.

awsKmsKeyDetails_creationDate :: Lens' AwsKmsKeyDetails (Maybe Double)

Indicates when the KMS key was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsKmsKeyDetails_description :: Lens' AwsKmsKeyDetails (Maybe Text)

A description of the key.