Amazonka.SecurityHub.Types.AwsOpenSearchServiceDomainClusterConfigDetails

Description

Documentation

data AwsOpenSearchServiceDomainClusterConfigDetails

Details about the configuration of an OpenSearch cluster.

See: newAwsOpenSearchServiceDomainClusterConfigDetails smart constructor.

Constructors

AwsOpenSearchServiceDomainClusterConfigDetails'

Fields

dedicatedMasterCount :: Maybe Int

The number of instances to use for the master node. If this attribute is specified, then DedicatedMasterEnabled must be true.

dedicatedMasterType :: Maybe Text

The hardware configuration of the computer that hosts the dedicated master node.

If this attribute is specified, then DedicatedMasterEnabled must be true.

dedicatedMasterEnabled :: Maybe Bool

Whether to use a dedicated master node for the OpenSearch domain. A dedicated master node performs cluster management tasks, but does not hold data or respond to data upload requests.

instanceCount :: Maybe Int

The number of data nodes to use in the OpenSearch domain.

zoneAwarenessEnabled :: Maybe Bool

Whether to enable zone awareness for the OpenSearch domain. When zone awareness is enabled, OpenSearch Service allocates the cluster's nodes and replica index shards across Availability Zones (AZs) in the same Region. This prevents data loss and minimizes downtime if a node or data center fails.

instanceType :: Maybe Text

The instance type for your data nodes.

warmEnabled :: Maybe Bool

Whether UltraWarm is enabled.

zoneAwarenessConfig :: Maybe AwsOpenSearchServiceDomainClusterConfigZoneAwarenessConfigDetails

Configuration options for zone awareness. Provided if ZoneAwarenessEnabled is true.

warmCount :: Maybe Int

The number of UltraWarm instances.

warmType :: Maybe Text

The type of UltraWarm instance.

newAwsOpenSearchServiceDomainClusterConfigDetails :: AwsOpenSearchServiceDomainClusterConfigDetails

Create a value of AwsOpenSearchServiceDomainClusterConfigDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dedicatedMasterCount:AwsOpenSearchServiceDomainClusterConfigDetails', awsOpenSearchServiceDomainClusterConfigDetails_dedicatedMasterCount - The number of instances to use for the master node. If this attribute is specified, then DedicatedMasterEnabled must be true.

$sel:dedicatedMasterType:AwsOpenSearchServiceDomainClusterConfigDetails', awsOpenSearchServiceDomainClusterConfigDetails_dedicatedMasterType - The hardware configuration of the computer that hosts the dedicated master node.

If this attribute is specified, then DedicatedMasterEnabled must be true.

$sel:dedicatedMasterEnabled:AwsOpenSearchServiceDomainClusterConfigDetails', awsOpenSearchServiceDomainClusterConfigDetails_dedicatedMasterEnabled - Whether to use a dedicated master node for the OpenSearch domain. A dedicated master node performs cluster management tasks, but does not hold data or respond to data upload requests.

$sel:instanceCount:AwsOpenSearchServiceDomainClusterConfigDetails', awsOpenSearchServiceDomainClusterConfigDetails_instanceCount - The number of data nodes to use in the OpenSearch domain.

$sel:zoneAwarenessEnabled:AwsOpenSearchServiceDomainClusterConfigDetails', awsOpenSearchServiceDomainClusterConfigDetails_zoneAwarenessEnabled - Whether to enable zone awareness for the OpenSearch domain. When zone awareness is enabled, OpenSearch Service allocates the cluster's nodes and replica index shards across Availability Zones (AZs) in the same Region. This prevents data loss and minimizes downtime if a node or data center fails.

$sel:instanceType:AwsOpenSearchServiceDomainClusterConfigDetails', awsOpenSearchServiceDomainClusterConfigDetails_instanceType - The instance type for your data nodes.

$sel:warmEnabled:AwsOpenSearchServiceDomainClusterConfigDetails', awsOpenSearchServiceDomainClusterConfigDetails_warmEnabled - Whether UltraWarm is enabled.

$sel:zoneAwarenessConfig:AwsOpenSearchServiceDomainClusterConfigDetails', awsOpenSearchServiceDomainClusterConfigDetails_zoneAwarenessConfig - Configuration options for zone awareness. Provided if ZoneAwarenessEnabled is true.

$sel:warmCount:AwsOpenSearchServiceDomainClusterConfigDetails', awsOpenSearchServiceDomainClusterConfigDetails_warmCount - The number of UltraWarm instances.

$sel:warmType:AwsOpenSearchServiceDomainClusterConfigDetails', awsOpenSearchServiceDomainClusterConfigDetails_warmType - The type of UltraWarm instance.

awsOpenSearchServiceDomainClusterConfigDetails_dedicatedMasterCount :: Lens' AwsOpenSearchServiceDomainClusterConfigDetails (Maybe Int)

The number of instances to use for the master node. If this attribute is specified, then DedicatedMasterEnabled must be true.

awsOpenSearchServiceDomainClusterConfigDetails_dedicatedMasterType :: Lens' AwsOpenSearchServiceDomainClusterConfigDetails (Maybe Text)

The hardware configuration of the computer that hosts the dedicated master node.

If this attribute is specified, then DedicatedMasterEnabled must be true.

awsOpenSearchServiceDomainClusterConfigDetails_dedicatedMasterEnabled :: Lens' AwsOpenSearchServiceDomainClusterConfigDetails (Maybe Bool)

Whether to use a dedicated master node for the OpenSearch domain. A dedicated master node performs cluster management tasks, but does not hold data or respond to data upload requests.

awsOpenSearchServiceDomainClusterConfigDetails_instanceCount :: Lens' AwsOpenSearchServiceDomainClusterConfigDetails (Maybe Int)

The number of data nodes to use in the OpenSearch domain.

awsOpenSearchServiceDomainClusterConfigDetails_zoneAwarenessEnabled :: Lens' AwsOpenSearchServiceDomainClusterConfigDetails (Maybe Bool)

Whether to enable zone awareness for the OpenSearch domain. When zone awareness is enabled, OpenSearch Service allocates the cluster's nodes and replica index shards across Availability Zones (AZs) in the same Region. This prevents data loss and minimizes downtime if a node or data center fails.

awsOpenSearchServiceDomainClusterConfigDetails_instanceType :: Lens' AwsOpenSearchServiceDomainClusterConfigDetails (Maybe Text)

The instance type for your data nodes.

awsOpenSearchServiceDomainClusterConfigDetails_warmEnabled :: Lens' AwsOpenSearchServiceDomainClusterConfigDetails (Maybe Bool)

Whether UltraWarm is enabled.

awsOpenSearchServiceDomainClusterConfigDetails_zoneAwarenessConfig :: Lens' AwsOpenSearchServiceDomainClusterConfigDetails (Maybe AwsOpenSearchServiceDomainClusterConfigZoneAwarenessConfigDetails)

Configuration options for zone awareness. Provided if ZoneAwarenessEnabled is true.

awsOpenSearchServiceDomainClusterConfigDetails_warmCount :: Lens' AwsOpenSearchServiceDomainClusterConfigDetails (Maybe Int)

The number of UltraWarm instances.

awsOpenSearchServiceDomainClusterConfigDetails_warmType :: Lens' AwsOpenSearchServiceDomainClusterConfigDetails (Maybe Text)

The type of UltraWarm instance.