Amazonka.SecurityHub.Types.AwsOpenSearchServiceDomainDetails

Description

Documentation

data AwsOpenSearchServiceDomainDetails

Information about an Amazon OpenSearch Service domain.

See: newAwsOpenSearchServiceDomainDetails smart constructor.

Constructors

AwsOpenSearchServiceDomainDetails'

Fields

engineVersion :: Maybe Text

The version of the domain engine.

nodeToNodeEncryptionOptions :: Maybe AwsOpenSearchServiceDomainNodeToNodeEncryptionOptionsDetails

Details about the configuration for node-to-node encryption.

accessPolicies :: Maybe Text

IAM policy document that specifies the access policies for the OpenSearch Service domain.

serviceSoftwareOptions :: Maybe AwsOpenSearchServiceDomainServiceSoftwareOptionsDetails

Information about the status of a domain relative to the latest service software.

arn :: Maybe Text

The ARN of the OpenSearch Service domain.

logPublishingOptions :: Maybe AwsOpenSearchServiceDomainLogPublishingOptionsDetails

Configures the CloudWatch Logs to publish for the OpenSearch domain.

clusterConfig :: Maybe AwsOpenSearchServiceDomainClusterConfigDetails

Details about the configuration of an OpenSearch cluster.

domainName :: Maybe Text

The name of the endpoint.

encryptionAtRestOptions :: Maybe AwsOpenSearchServiceDomainEncryptionAtRestOptionsDetails

Details about the configuration for encryption at rest.

vpcOptions :: Maybe AwsOpenSearchServiceDomainVpcOptionsDetails

Information that OpenSearch Service derives based on VPCOptions for the domain.

id :: Maybe Text

The identifier of the domain.

domainEndpoints :: Maybe (HashMap Text Text)

The domain endpoints. Used if the OpenSearch domain resides in a VPC.

This is a map of key-value pairs. The key is always vpc. The value is the endpoint.

domainEndpoint :: Maybe Text

The domain endpoint.

domainEndpointOptions :: Maybe AwsOpenSearchServiceDomainDomainEndpointOptionsDetails

Additional options for the domain endpoint.

newAwsOpenSearchServiceDomainDetails :: AwsOpenSearchServiceDomainDetails

Create a value of AwsOpenSearchServiceDomainDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_engineVersion - The version of the domain engine.

$sel:nodeToNodeEncryptionOptions:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_nodeToNodeEncryptionOptions - Details about the configuration for node-to-node encryption.

$sel:accessPolicies:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_accessPolicies - IAM policy document that specifies the access policies for the OpenSearch Service domain.

$sel:serviceSoftwareOptions:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_serviceSoftwareOptions - Information about the status of a domain relative to the latest service software.

$sel:arn:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_arn - The ARN of the OpenSearch Service domain.

$sel:logPublishingOptions:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_logPublishingOptions - Configures the CloudWatch Logs to publish for the OpenSearch domain.

$sel:clusterConfig:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_clusterConfig - Details about the configuration of an OpenSearch cluster.

$sel:domainName:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_domainName - The name of the endpoint.

$sel:encryptionAtRestOptions:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_encryptionAtRestOptions - Details about the configuration for encryption at rest.

$sel:vpcOptions:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_vpcOptions - Information that OpenSearch Service derives based on VPCOptions for the domain.

$sel:id:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_id - The identifier of the domain.

$sel:domainEndpoints:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_domainEndpoints - The domain endpoints. Used if the OpenSearch domain resides in a VPC.

This is a map of key-value pairs. The key is always vpc. The value is the endpoint.

$sel:domainEndpoint:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_domainEndpoint - The domain endpoint.

$sel:domainEndpointOptions:AwsOpenSearchServiceDomainDetails', awsOpenSearchServiceDomainDetails_domainEndpointOptions - Additional options for the domain endpoint.

awsOpenSearchServiceDomainDetails_engineVersion :: Lens' AwsOpenSearchServiceDomainDetails (Maybe Text)

The version of the domain engine.

awsOpenSearchServiceDomainDetails_nodeToNodeEncryptionOptions :: Lens' AwsOpenSearchServiceDomainDetails (Maybe AwsOpenSearchServiceDomainNodeToNodeEncryptionOptionsDetails)

Details about the configuration for node-to-node encryption.

awsOpenSearchServiceDomainDetails_accessPolicies :: Lens' AwsOpenSearchServiceDomainDetails (Maybe Text)

IAM policy document that specifies the access policies for the OpenSearch Service domain.

awsOpenSearchServiceDomainDetails_serviceSoftwareOptions :: Lens' AwsOpenSearchServiceDomainDetails (Maybe AwsOpenSearchServiceDomainServiceSoftwareOptionsDetails)

Information about the status of a domain relative to the latest service software.

awsOpenSearchServiceDomainDetails_arn :: Lens' AwsOpenSearchServiceDomainDetails (Maybe Text)

The ARN of the OpenSearch Service domain.

awsOpenSearchServiceDomainDetails_logPublishingOptions :: Lens' AwsOpenSearchServiceDomainDetails (Maybe AwsOpenSearchServiceDomainLogPublishingOptionsDetails)

Configures the CloudWatch Logs to publish for the OpenSearch domain.

awsOpenSearchServiceDomainDetails_clusterConfig :: Lens' AwsOpenSearchServiceDomainDetails (Maybe AwsOpenSearchServiceDomainClusterConfigDetails)

Details about the configuration of an OpenSearch cluster.

awsOpenSearchServiceDomainDetails_domainName :: Lens' AwsOpenSearchServiceDomainDetails (Maybe Text)

The name of the endpoint.

awsOpenSearchServiceDomainDetails_encryptionAtRestOptions :: Lens' AwsOpenSearchServiceDomainDetails (Maybe AwsOpenSearchServiceDomainEncryptionAtRestOptionsDetails)

Details about the configuration for encryption at rest.

awsOpenSearchServiceDomainDetails_vpcOptions :: Lens' AwsOpenSearchServiceDomainDetails (Maybe AwsOpenSearchServiceDomainVpcOptionsDetails)

Information that OpenSearch Service derives based on VPCOptions for the domain.

awsOpenSearchServiceDomainDetails_id :: Lens' AwsOpenSearchServiceDomainDetails (Maybe Text)

The identifier of the domain.

awsOpenSearchServiceDomainDetails_domainEndpoints :: Lens' AwsOpenSearchServiceDomainDetails (Maybe (HashMap Text Text))

The domain endpoints. Used if the OpenSearch domain resides in a VPC.

This is a map of key-value pairs. The key is always vpc. The value is the endpoint.

awsOpenSearchServiceDomainDetails_domainEndpoint :: Lens' AwsOpenSearchServiceDomainDetails (Maybe Text)

The domain endpoint.

awsOpenSearchServiceDomainDetails_domainEndpointOptions :: Lens' AwsOpenSearchServiceDomainDetails (Maybe AwsOpenSearchServiceDomainDomainEndpointOptionsDetails)

Additional options for the domain endpoint.