Amazonka.SecurityHub.Types.AwsOpenSearchServiceDomainDomainEndpointOptionsDetails

Description

Documentation

data AwsOpenSearchServiceDomainDomainEndpointOptionsDetails

Information about additional options for the domain endpoint.

See: newAwsOpenSearchServiceDomainDomainEndpointOptionsDetails smart constructor.

Constructors

AwsOpenSearchServiceDomainDomainEndpointOptionsDetails'

Fields

enforceHTTPS :: Maybe Bool

Whether to require that all traffic to the domain arrive over HTTPS.

tLSSecurityPolicy :: Maybe Text

The TLS security policy to apply to the HTTPS endpoint of the OpenSearch domain.

customEndpointEnabled :: Maybe Bool

Whether to enable a custom endpoint for the domain.

customEndpoint :: Maybe Text

The fully qualified URL for the custom endpoint.

customEndpointCertificateArn :: Maybe Text

The ARN for the security certificate. The certificate is managed in ACM.

newAwsOpenSearchServiceDomainDomainEndpointOptionsDetails :: AwsOpenSearchServiceDomainDomainEndpointOptionsDetails

Create a value of AwsOpenSearchServiceDomainDomainEndpointOptionsDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enforceHTTPS:AwsOpenSearchServiceDomainDomainEndpointOptionsDetails', awsOpenSearchServiceDomainDomainEndpointOptionsDetails_enforceHTTPS - Whether to require that all traffic to the domain arrive over HTTPS.

$sel:tLSSecurityPolicy:AwsOpenSearchServiceDomainDomainEndpointOptionsDetails', awsOpenSearchServiceDomainDomainEndpointOptionsDetails_tLSSecurityPolicy - The TLS security policy to apply to the HTTPS endpoint of the OpenSearch domain.

$sel:customEndpointEnabled:AwsOpenSearchServiceDomainDomainEndpointOptionsDetails', awsOpenSearchServiceDomainDomainEndpointOptionsDetails_customEndpointEnabled - Whether to enable a custom endpoint for the domain.

$sel:customEndpoint:AwsOpenSearchServiceDomainDomainEndpointOptionsDetails', awsOpenSearchServiceDomainDomainEndpointOptionsDetails_customEndpoint - The fully qualified URL for the custom endpoint.

$sel:customEndpointCertificateArn:AwsOpenSearchServiceDomainDomainEndpointOptionsDetails', awsOpenSearchServiceDomainDomainEndpointOptionsDetails_customEndpointCertificateArn - The ARN for the security certificate. The certificate is managed in ACM.

awsOpenSearchServiceDomainDomainEndpointOptionsDetails_enforceHTTPS :: Lens' AwsOpenSearchServiceDomainDomainEndpointOptionsDetails (Maybe Bool)

Whether to require that all traffic to the domain arrive over HTTPS.

awsOpenSearchServiceDomainDomainEndpointOptionsDetails_tLSSecurityPolicy :: Lens' AwsOpenSearchServiceDomainDomainEndpointOptionsDetails (Maybe Text)

The TLS security policy to apply to the HTTPS endpoint of the OpenSearch domain.

awsOpenSearchServiceDomainDomainEndpointOptionsDetails_customEndpointEnabled :: Lens' AwsOpenSearchServiceDomainDomainEndpointOptionsDetails (Maybe Bool)

Whether to enable a custom endpoint for the domain.

awsOpenSearchServiceDomainDomainEndpointOptionsDetails_customEndpoint :: Lens' AwsOpenSearchServiceDomainDomainEndpointOptionsDetails (Maybe Text)

The fully qualified URL for the custom endpoint.

awsOpenSearchServiceDomainDomainEndpointOptionsDetails_customEndpointCertificateArn :: Lens' AwsOpenSearchServiceDomainDomainEndpointOptionsDetails (Maybe Text)

The ARN for the security certificate. The certificate is managed in ACM.