Amazonka.SecurityHub.Types.AwsRdsDbClusterDetails

Description

Documentation

data AwsRdsDbClusterDetails

Information about an Amazon RDS DB cluster.

See: newAwsRdsDbClusterDetails smart constructor.

Constructors

AwsRdsDbClusterDetails'

Fields

engineVersion :: Maybe Text

The version number of the database engine to use.

status :: Maybe Text

The current status of this DB cluster.

dbClusterMembers :: Maybe [AwsRdsDbClusterMember]

The list of instances that make up the DB cluster.

deletionProtection :: Maybe Bool

Whether the DB cluster has deletion protection enabled.

storageEncrypted :: Maybe Bool

Whether the DB cluster is encrypted.

dbClusterIdentifier :: Maybe Text

The DB cluster identifier that the user assigned to the cluster. This identifier is the unique key that identifies a DB cluster.

readReplicaIdentifiers :: Maybe [Text]

The identifiers of the read replicas that are associated with this DB cluster.

dbClusterParameterGroup :: Maybe Text

The name of the DB cluster parameter group for the DB cluster.

hostedZoneId :: Maybe Text

Specifies the identifier that Amazon Route 53 assigns when you create a hosted zone.

masterUsername :: Maybe Text

The name of the master user for the DB cluster.

iamDatabaseAuthenticationEnabled :: Maybe Bool

Whether the mapping of IAM accounts to database accounts is enabled.

dbClusterResourceId :: Maybe Text

The identifier of the DB cluster. The identifier must be unique within each Amazon Web Services Region and is immutable.

customEndpoints :: Maybe [Text]

A list of custom endpoints for the DB cluster.

engine :: Maybe Text

The name of the database engine to use for this DB cluster.

httpEndpointEnabled :: Maybe Bool

Whether the HTTP endpoint for an Aurora Serverless DB cluster is enabled.

crossAccountClone :: Maybe Bool

Whether the DB cluster is a clone of a DB cluster owned by a different Amazon Web Services account.

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Uses the format <day>:HH:MM-<day>:HH:MM.

For the day values, use mon|tue|wed|thu|fri|sat|sun.

For example, sun:09:32-sun:10:02.

availabilityZones :: Maybe [Text]

A list of Availability Zones (AZs) where instances in the DB cluster can be created.

kmsKeyId :: Maybe Text

The ARN of the KMS master key that is used to encrypt the database instances in the DB cluster.

preferredBackupWindow :: Maybe Text

The range of time each day when automated backups are created, if automated backups are enabled.

Uses the format HH:MM-HH:MM. For example, 04:52-05:22.

associatedRoles :: Maybe [AwsRdsDbClusterAssociatedRole]

A list of the IAM roles that are associated with the DB cluster.

vpcSecurityGroups :: Maybe [AwsRdsDbInstanceVpcSecurityGroup]

A list of VPC security groups that the DB cluster belongs to.

backupRetentionPeriod :: Maybe Int

The number of days for which automated backups are retained.

dbSubnetGroup :: Maybe Text

The subnet group that is associated with the DB cluster, including the name, description, and subnets in the subnet group.

multiAz :: Maybe Bool

Whether the DB cluster has instances in multiple Availability Zones.

databaseName :: Maybe Text

The name of the database.

engineMode :: Maybe Text

The database engine mode of the DB cluster.

enabledCloudWatchLogsExports :: Maybe [Text]

A list of log types that this DB cluster is configured to export to CloudWatch Logs.

activityStreamStatus :: Maybe Text

The status of the database activity stream.

allocatedStorage :: Maybe Int

For all database engines except Aurora, specifies the allocated storage size in gibibytes (GiB).

copyTagsToSnapshot :: Maybe Bool

Whether tags are copied from the DB cluster to snapshots of the DB cluster.

clusterCreateTime :: Maybe Text

Indicates when the DB cluster was created, in Universal Coordinated Time (UTC).

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

endpoint :: Maybe Text

The connection endpoint for the primary instance of the DB cluster.

readerEndpoint :: Maybe Text

The reader endpoint for the DB cluster.

dbClusterOptionGroupMemberships :: Maybe [AwsRdsDbClusterOptionGroupMembership]

The list of option group memberships for this DB cluster.

port :: Maybe Int

The port number on which the DB instances in the DB cluster accept connections.

domainMemberships :: Maybe [AwsRdsDbDomainMembership]

The Active Directory domain membership records that are associated with the DB cluster.

newAwsRdsDbClusterDetails :: AwsRdsDbClusterDetails

Create a value of AwsRdsDbClusterDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_engineVersion - The version number of the database engine to use.

$sel:status:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_status - The current status of this DB cluster.

$sel:dbClusterMembers:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_dbClusterMembers - The list of instances that make up the DB cluster.

$sel:deletionProtection:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_deletionProtection - Whether the DB cluster has deletion protection enabled.

$sel:storageEncrypted:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_storageEncrypted - Whether the DB cluster is encrypted.

$sel:dbClusterIdentifier:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_dbClusterIdentifier - The DB cluster identifier that the user assigned to the cluster. This identifier is the unique key that identifies a DB cluster.

$sel:readReplicaIdentifiers:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_readReplicaIdentifiers - The identifiers of the read replicas that are associated with this DB cluster.

$sel:dbClusterParameterGroup:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_dbClusterParameterGroup - The name of the DB cluster parameter group for the DB cluster.

$sel:hostedZoneId:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_hostedZoneId - Specifies the identifier that Amazon Route 53 assigns when you create a hosted zone.

$sel:masterUsername:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_masterUsername - The name of the master user for the DB cluster.

$sel:iamDatabaseAuthenticationEnabled:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_iamDatabaseAuthenticationEnabled - Whether the mapping of IAM accounts to database accounts is enabled.

$sel:dbClusterResourceId:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_dbClusterResourceId - The identifier of the DB cluster. The identifier must be unique within each Amazon Web Services Region and is immutable.

$sel:customEndpoints:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_customEndpoints - A list of custom endpoints for the DB cluster.

$sel:engine:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_engine - The name of the database engine to use for this DB cluster.

$sel:httpEndpointEnabled:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_httpEndpointEnabled - Whether the HTTP endpoint for an Aurora Serverless DB cluster is enabled.

$sel:crossAccountClone:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_crossAccountClone - Whether the DB cluster is a clone of a DB cluster owned by a different Amazon Web Services account.

$sel:preferredMaintenanceWindow:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Uses the format <day>:HH:MM-<day>:HH:MM.

For the day values, use mon|tue|wed|thu|fri|sat|sun.

For example, sun:09:32-sun:10:02.

$sel:availabilityZones:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_availabilityZones - A list of Availability Zones (AZs) where instances in the DB cluster can be created.

$sel:kmsKeyId:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_kmsKeyId - The ARN of the KMS master key that is used to encrypt the database instances in the DB cluster.

$sel:preferredBackupWindow:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_preferredBackupWindow - The range of time each day when automated backups are created, if automated backups are enabled.

Uses the format HH:MM-HH:MM. For example, 04:52-05:22.

$sel:associatedRoles:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_associatedRoles - A list of the IAM roles that are associated with the DB cluster.

$sel:vpcSecurityGroups:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_vpcSecurityGroups - A list of VPC security groups that the DB cluster belongs to.

$sel:backupRetentionPeriod:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_backupRetentionPeriod - The number of days for which automated backups are retained.

$sel:dbSubnetGroup:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_dbSubnetGroup - The subnet group that is associated with the DB cluster, including the name, description, and subnets in the subnet group.

$sel:multiAz:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_multiAz - Whether the DB cluster has instances in multiple Availability Zones.

$sel:databaseName:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_databaseName - The name of the database.

$sel:engineMode:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_engineMode - The database engine mode of the DB cluster.

$sel:enabledCloudWatchLogsExports:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_enabledCloudWatchLogsExports - A list of log types that this DB cluster is configured to export to CloudWatch Logs.

$sel:activityStreamStatus:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_activityStreamStatus - The status of the database activity stream.

$sel:allocatedStorage:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_allocatedStorage - For all database engines except Aurora, specifies the allocated storage size in gibibytes (GiB).

$sel:copyTagsToSnapshot:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_copyTagsToSnapshot - Whether tags are copied from the DB cluster to snapshots of the DB cluster.

$sel:clusterCreateTime:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_clusterCreateTime - Indicates when the DB cluster was created, in Universal Coordinated Time (UTC).

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:endpoint:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_endpoint - The connection endpoint for the primary instance of the DB cluster.

$sel:readerEndpoint:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_readerEndpoint - The reader endpoint for the DB cluster.

$sel:dbClusterOptionGroupMemberships:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_dbClusterOptionGroupMemberships - The list of option group memberships for this DB cluster.

$sel:port:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_port - The port number on which the DB instances in the DB cluster accept connections.

$sel:domainMemberships:AwsRdsDbClusterDetails', awsRdsDbClusterDetails_domainMemberships - The Active Directory domain membership records that are associated with the DB cluster.

awsRdsDbClusterDetails_engineVersion :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The version number of the database engine to use.

awsRdsDbClusterDetails_status :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The current status of this DB cluster.

awsRdsDbClusterDetails_dbClusterMembers :: Lens' AwsRdsDbClusterDetails (Maybe [AwsRdsDbClusterMember])

The list of instances that make up the DB cluster.

awsRdsDbClusterDetails_deletionProtection :: Lens' AwsRdsDbClusterDetails (Maybe Bool)

Whether the DB cluster has deletion protection enabled.

awsRdsDbClusterDetails_storageEncrypted :: Lens' AwsRdsDbClusterDetails (Maybe Bool)

Whether the DB cluster is encrypted.

awsRdsDbClusterDetails_dbClusterIdentifier :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The DB cluster identifier that the user assigned to the cluster. This identifier is the unique key that identifies a DB cluster.

awsRdsDbClusterDetails_readReplicaIdentifiers :: Lens' AwsRdsDbClusterDetails (Maybe [Text])

The identifiers of the read replicas that are associated with this DB cluster.

awsRdsDbClusterDetails_dbClusterParameterGroup :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The name of the DB cluster parameter group for the DB cluster.

awsRdsDbClusterDetails_hostedZoneId :: Lens' AwsRdsDbClusterDetails (Maybe Text)

Specifies the identifier that Amazon Route 53 assigns when you create a hosted zone.

awsRdsDbClusterDetails_masterUsername :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The name of the master user for the DB cluster.

awsRdsDbClusterDetails_iamDatabaseAuthenticationEnabled :: Lens' AwsRdsDbClusterDetails (Maybe Bool)

Whether the mapping of IAM accounts to database accounts is enabled.

awsRdsDbClusterDetails_dbClusterResourceId :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The identifier of the DB cluster. The identifier must be unique within each Amazon Web Services Region and is immutable.

awsRdsDbClusterDetails_customEndpoints :: Lens' AwsRdsDbClusterDetails (Maybe [Text])

A list of custom endpoints for the DB cluster.

awsRdsDbClusterDetails_engine :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The name of the database engine to use for this DB cluster.

awsRdsDbClusterDetails_httpEndpointEnabled :: Lens' AwsRdsDbClusterDetails (Maybe Bool)

Whether the HTTP endpoint for an Aurora Serverless DB cluster is enabled.

awsRdsDbClusterDetails_crossAccountClone :: Lens' AwsRdsDbClusterDetails (Maybe Bool)

Whether the DB cluster is a clone of a DB cluster owned by a different Amazon Web Services account.

awsRdsDbClusterDetails_preferredMaintenanceWindow :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Uses the format <day>:HH:MM-<day>:HH:MM.

For the day values, use mon|tue|wed|thu|fri|sat|sun.

For example, sun:09:32-sun:10:02.

awsRdsDbClusterDetails_availabilityZones :: Lens' AwsRdsDbClusterDetails (Maybe [Text])

A list of Availability Zones (AZs) where instances in the DB cluster can be created.

awsRdsDbClusterDetails_kmsKeyId :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The ARN of the KMS master key that is used to encrypt the database instances in the DB cluster.

awsRdsDbClusterDetails_preferredBackupWindow :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The range of time each day when automated backups are created, if automated backups are enabled.

Uses the format HH:MM-HH:MM. For example, 04:52-05:22.

awsRdsDbClusterDetails_associatedRoles :: Lens' AwsRdsDbClusterDetails (Maybe [AwsRdsDbClusterAssociatedRole])

A list of the IAM roles that are associated with the DB cluster.

awsRdsDbClusterDetails_vpcSecurityGroups :: Lens' AwsRdsDbClusterDetails (Maybe [AwsRdsDbInstanceVpcSecurityGroup])

A list of VPC security groups that the DB cluster belongs to.

awsRdsDbClusterDetails_backupRetentionPeriod :: Lens' AwsRdsDbClusterDetails (Maybe Int)

The number of days for which automated backups are retained.

awsRdsDbClusterDetails_dbSubnetGroup :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The subnet group that is associated with the DB cluster, including the name, description, and subnets in the subnet group.

awsRdsDbClusterDetails_multiAz :: Lens' AwsRdsDbClusterDetails (Maybe Bool)

Whether the DB cluster has instances in multiple Availability Zones.

awsRdsDbClusterDetails_databaseName :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The name of the database.

awsRdsDbClusterDetails_engineMode :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The database engine mode of the DB cluster.

awsRdsDbClusterDetails_enabledCloudWatchLogsExports :: Lens' AwsRdsDbClusterDetails (Maybe [Text])

A list of log types that this DB cluster is configured to export to CloudWatch Logs.

awsRdsDbClusterDetails_activityStreamStatus :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The status of the database activity stream.

awsRdsDbClusterDetails_allocatedStorage :: Lens' AwsRdsDbClusterDetails (Maybe Int)

For all database engines except Aurora, specifies the allocated storage size in gibibytes (GiB).

awsRdsDbClusterDetails_copyTagsToSnapshot :: Lens' AwsRdsDbClusterDetails (Maybe Bool)

Whether tags are copied from the DB cluster to snapshots of the DB cluster.

awsRdsDbClusterDetails_clusterCreateTime :: Lens' AwsRdsDbClusterDetails (Maybe Text)

Indicates when the DB cluster was created, in Universal Coordinated Time (UTC).

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsRdsDbClusterDetails_endpoint :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The connection endpoint for the primary instance of the DB cluster.

awsRdsDbClusterDetails_readerEndpoint :: Lens' AwsRdsDbClusterDetails (Maybe Text)

The reader endpoint for the DB cluster.

awsRdsDbClusterDetails_dbClusterOptionGroupMemberships :: Lens' AwsRdsDbClusterDetails (Maybe [AwsRdsDbClusterOptionGroupMembership])

The list of option group memberships for this DB cluster.

awsRdsDbClusterDetails_port :: Lens' AwsRdsDbClusterDetails (Maybe Int)

The port number on which the DB instances in the DB cluster accept connections.

awsRdsDbClusterDetails_domainMemberships :: Lens' AwsRdsDbClusterDetails (Maybe [AwsRdsDbDomainMembership])

The Active Directory domain membership records that are associated with the DB cluster.