Amazonka.SecurityHub.Types.AwsRdsDbInstanceDetails

Description

Documentation

data AwsRdsDbInstanceDetails

Contains the details of an Amazon RDS DB instance.

See: newAwsRdsDbInstanceDetails smart constructor.

Constructors

AwsRdsDbInstanceDetails'

Fields

dbSecurityGroups :: Maybe [Text]

A list of the DB security groups to assign to the DB instance.

engineVersion :: Maybe Text

Indicates the database engine version.

deletionProtection :: Maybe Bool

Indicates whether the DB instance has deletion protection enabled.

When deletion protection is enabled, the database cannot be deleted.

storageEncrypted :: Maybe Bool

Specifies whether the DB instance is encrypted.

dbClusterIdentifier :: Maybe Text

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

publiclyAccessible :: Maybe Bool

Specifies the accessibility options for the DB instance.

A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address.

A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

autoMinorVersionUpgrade :: Maybe Bool

Indicates whether minor version patches are applied automatically.

masterUsername :: Maybe Text

The master user name of the DB instance.

readReplicaDBInstanceIdentifiers :: Maybe [Text]

List of identifiers of the read replicas associated with this DB instance.

iAMDatabaseAuthenticationEnabled :: Maybe Bool

True if mapping of IAM accounts to database accounts is enabled, and otherwise false.

IAM database authentication can be enabled for the following database engines.

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher

monitoringRoleArn :: Maybe Text

The ARN for the IAM role that permits Amazon RDS to send enhanced monitoring metrics to CloudWatch Logs.

iops :: Maybe Int

Specifies the provisioned IOPS (I/O operations per second) for this DB instance.

instanceCreateTime :: Maybe Text

Indicates when the DB instance was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

readReplicaSourceDBInstanceIdentifier :: Maybe Text

If this DB instance is a read replica, contains the identifier of the source DB instance.

monitoringInterval :: Maybe Int

The interval, in seconds, between points when enhanced monitoring metrics are collected for the DB instance.

engine :: Maybe Text

Provides the name of the database engine to use for this DB instance.

processorFeatures :: Maybe [AwsRdsDbProcessorFeature]

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

latestRestorableTime :: Maybe Text

Specifies the latest time to which a database can be restored with point-in-time restore.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

dbInstanceClass :: Maybe Text

Contains the name of the compute and memory capacity class of the DB instance.

promotionTier :: Maybe Int

The order in which to promote an Aurora replica to the primary instance after a failure of the existing primary instance.

licenseModel :: Maybe Text

License model information for this DB instance.

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Uses the format <day>:HH:MM-<day>:HH:MM.

For the day values, use mon|tue|wed|thu|fri|sat|sun.

For example, sun:09:32-sun:10:02.

performanceInsightsRetentionPeriod :: Maybe Int

The number of days to retain Performance Insights data.

cACertificateIdentifier :: Maybe Text

The identifier of the CA certificate for this DB instance.

dbInstanceIdentifier :: Maybe Text

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

characterSetName :: Maybe Text

The name of the character set that this DB instance is associated with.

maxAllocatedStorage :: Maybe Int

The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

kmsKeyId :: Maybe Text

If StorageEncrypted is true, the KMS key identifier for the encrypted DB instance.

preferredBackupWindow :: Maybe Text

The range of time each day when automated backups are created, if automated backups are enabled.

Uses the format HH:MM-HH:MM. For example, 04:52-05:22.

associatedRoles :: Maybe [AwsRdsDbInstanceAssociatedRole]

The IAM roles associated with the DB instance.

availabilityZone :: Maybe Text

The Availability Zone where the DB instance will be created.

vpcSecurityGroups :: Maybe [AwsRdsDbInstanceVpcSecurityGroup]

A list of VPC security groups that the DB instance belongs to.

backupRetentionPeriod :: Maybe Int

The number of days for which to retain automated backups.

performanceInsightsKmsKeyId :: Maybe Text

The identifier of the KMS key used to encrypt the Performance Insights data.

dbSubnetGroup :: Maybe AwsRdsDbSubnetGroup

Information about the subnet group that is associated with the DB instance.

multiAz :: Maybe Bool

Whether the DB instance is a multiple Availability Zone deployment.

listenerEndpoint :: Maybe AwsRdsDbInstanceEndpoint

optionGroupMemberships :: Maybe [AwsRdsDbOptionGroupMembership]

The list of option group memberships for this DB instance.

enhancedMonitoringResourceArn :: Maybe Text

The ARN of the CloudWatch Logs log stream that receives the enhanced monitoring metrics data for the DB instance.

secondaryAvailabilityZone :: Maybe Text

For a DB instance with multi-Availability Zone support, the name of the secondary Availability Zone.

enabledCloudWatchLogsExports :: Maybe [Text]

A list of log types that this DB instance is configured to export to CloudWatch Logs.

performanceInsightsEnabled :: Maybe Bool

Indicates whether Performance Insights is enabled for the DB instance.

allocatedStorage :: Maybe Int

The amount of storage (in gigabytes) to initially allocate for the DB instance.

dbiResourceId :: Maybe Text

The Amazon Web Services Region-unique, immutable identifier for the DB instance. This identifier is found in CloudTrail log entries whenever the KMS key for the DB instance is accessed.

dbInstanceStatus :: Maybe Text

The current status of the DB instance.

copyTagsToSnapshot :: Maybe Bool

Whether to copy resource tags to snapshots of the DB instance.

timezone :: Maybe Text

The time zone of the DB instance.

tdeCredentialArn :: Maybe Text

The ARN from the key store with which the instance is associated for TDE encryption.

dbParameterGroups :: Maybe [AwsRdsDbParameterGroup]

A list of the DB parameter groups to assign to the DB instance.

endpoint :: Maybe AwsRdsDbInstanceEndpoint

Specifies the connection endpoint.

dbInstancePort :: Maybe Int

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

pendingModifiedValues :: Maybe AwsRdsDbPendingModifiedValues

Changes to the DB instance that are currently pending.

readReplicaDBClusterIdentifiers :: Maybe [Text]

List of identifiers of Aurora DB clusters to which the RDS DB instance is replicated as a read replica.

storageType :: Maybe Text

The storage type for the DB instance.

statusInfos :: Maybe [AwsRdsDbStatusInfo]

The status of a read replica. If the instance isn't a read replica, this is empty.

domainMemberships :: Maybe [AwsRdsDbDomainMembership]

The Active Directory domain membership records associated with the DB instance.

dbName :: Maybe Text

The meaning of this parameter differs according to the database engine you use.

MySQL, MariaDB, SQL Server, PostgreSQL

Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance.

Oracle

Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.

newAwsRdsDbInstanceDetails :: AwsRdsDbInstanceDetails

Create a value of AwsRdsDbInstanceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSecurityGroups:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_dbSecurityGroups - A list of the DB security groups to assign to the DB instance.

$sel:engineVersion:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_engineVersion - Indicates the database engine version.

$sel:deletionProtection:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_deletionProtection - Indicates whether the DB instance has deletion protection enabled.

When deletion protection is enabled, the database cannot be deleted.

$sel:storageEncrypted:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_storageEncrypted - Specifies whether the DB instance is encrypted.

$sel:dbClusterIdentifier:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_dbClusterIdentifier - If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

$sel:publiclyAccessible:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_publiclyAccessible - Specifies the accessibility options for the DB instance.

A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address.

A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

$sel:autoMinorVersionUpgrade:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_autoMinorVersionUpgrade - Indicates whether minor version patches are applied automatically.

$sel:masterUsername:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_masterUsername - The master user name of the DB instance.

$sel:readReplicaDBInstanceIdentifiers:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_readReplicaDBInstanceIdentifiers - List of identifiers of the read replicas associated with this DB instance.

$sel:iAMDatabaseAuthenticationEnabled:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_iAMDatabaseAuthenticationEnabled - True if mapping of IAM accounts to database accounts is enabled, and otherwise false.

IAM database authentication can be enabled for the following database engines.

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher

$sel:monitoringRoleArn:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_monitoringRoleArn - The ARN for the IAM role that permits Amazon RDS to send enhanced monitoring metrics to CloudWatch Logs.

$sel:iops:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_iops - Specifies the provisioned IOPS (I/O operations per second) for this DB instance.

$sel:instanceCreateTime:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_instanceCreateTime - Indicates when the DB instance was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:readReplicaSourceDBInstanceIdentifier:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_readReplicaSourceDBInstanceIdentifier - If this DB instance is a read replica, contains the identifier of the source DB instance.

$sel:monitoringInterval:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_monitoringInterval - The interval, in seconds, between points when enhanced monitoring metrics are collected for the DB instance.

$sel:engine:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_engine - Provides the name of the database engine to use for this DB instance.

$sel:processorFeatures:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_processorFeatures - The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

$sel:latestRestorableTime:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_latestRestorableTime - Specifies the latest time to which a database can be restored with point-in-time restore.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:dbInstanceClass:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_dbInstanceClass - Contains the name of the compute and memory capacity class of the DB instance.

$sel:promotionTier:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_promotionTier - The order in which to promote an Aurora replica to the primary instance after a failure of the existing primary instance.

$sel:licenseModel:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_licenseModel - License model information for this DB instance.

$sel:preferredMaintenanceWindow:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Uses the format <day>:HH:MM-<day>:HH:MM.

For the day values, use mon|tue|wed|thu|fri|sat|sun.

For example, sun:09:32-sun:10:02.

$sel:performanceInsightsRetentionPeriod:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_performanceInsightsRetentionPeriod - The number of days to retain Performance Insights data.

$sel:cACertificateIdentifier:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_cACertificateIdentifier - The identifier of the CA certificate for this DB instance.

$sel:dbInstanceIdentifier:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_dbInstanceIdentifier - Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

$sel:characterSetName:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_characterSetName - The name of the character set that this DB instance is associated with.

$sel:maxAllocatedStorage:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_maxAllocatedStorage - The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

$sel:kmsKeyId:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_kmsKeyId - If StorageEncrypted is true, the KMS key identifier for the encrypted DB instance.

$sel:preferredBackupWindow:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_preferredBackupWindow - The range of time each day when automated backups are created, if automated backups are enabled.

Uses the format HH:MM-HH:MM. For example, 04:52-05:22.

$sel:associatedRoles:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_associatedRoles - The IAM roles associated with the DB instance.

$sel:availabilityZone:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_availabilityZone - The Availability Zone where the DB instance will be created.

$sel:vpcSecurityGroups:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_vpcSecurityGroups - A list of VPC security groups that the DB instance belongs to.

$sel:backupRetentionPeriod:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_backupRetentionPeriod - The number of days for which to retain automated backups.

$sel:performanceInsightsKmsKeyId:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_performanceInsightsKmsKeyId - The identifier of the KMS key used to encrypt the Performance Insights data.

$sel:dbSubnetGroup:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_dbSubnetGroup - Information about the subnet group that is associated with the DB instance.

$sel:multiAz:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_multiAz - Whether the DB instance is a multiple Availability Zone deployment.

$sel:listenerEndpoint:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_listenerEndpoint - Undocumented member.

$sel:optionGroupMemberships:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_optionGroupMemberships - The list of option group memberships for this DB instance.

$sel:enhancedMonitoringResourceArn:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_enhancedMonitoringResourceArn - The ARN of the CloudWatch Logs log stream that receives the enhanced monitoring metrics data for the DB instance.

$sel:secondaryAvailabilityZone:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_secondaryAvailabilityZone - For a DB instance with multi-Availability Zone support, the name of the secondary Availability Zone.

$sel:enabledCloudWatchLogsExports:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_enabledCloudWatchLogsExports - A list of log types that this DB instance is configured to export to CloudWatch Logs.

$sel:performanceInsightsEnabled:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_performanceInsightsEnabled - Indicates whether Performance Insights is enabled for the DB instance.

$sel:allocatedStorage:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_allocatedStorage - The amount of storage (in gigabytes) to initially allocate for the DB instance.

$sel:dbiResourceId:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_dbiResourceId - The Amazon Web Services Region-unique, immutable identifier for the DB instance. This identifier is found in CloudTrail log entries whenever the KMS key for the DB instance is accessed.

$sel:dbInstanceStatus:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_dbInstanceStatus - The current status of the DB instance.

$sel:copyTagsToSnapshot:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_copyTagsToSnapshot - Whether to copy resource tags to snapshots of the DB instance.

$sel:timezone:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_timezone - The time zone of the DB instance.

$sel:tdeCredentialArn:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_tdeCredentialArn - The ARN from the key store with which the instance is associated for TDE encryption.

$sel:dbParameterGroups:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_dbParameterGroups - A list of the DB parameter groups to assign to the DB instance.

$sel:endpoint:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_endpoint - Specifies the connection endpoint.

$sel:dbInstancePort:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_dbInstancePort - Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

$sel:pendingModifiedValues:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_pendingModifiedValues - Changes to the DB instance that are currently pending.

$sel:readReplicaDBClusterIdentifiers:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_readReplicaDBClusterIdentifiers - List of identifiers of Aurora DB clusters to which the RDS DB instance is replicated as a read replica.

$sel:storageType:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_storageType - The storage type for the DB instance.

$sel:statusInfos:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_statusInfos - The status of a read replica. If the instance isn't a read replica, this is empty.

$sel:domainMemberships:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_domainMemberships - The Active Directory domain membership records associated with the DB instance.

$sel:dbName:AwsRdsDbInstanceDetails', awsRdsDbInstanceDetails_dbName - The meaning of this parameter differs according to the database engine you use.

MySQL, MariaDB, SQL Server, PostgreSQL

Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance.

Oracle

Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.

awsRdsDbInstanceDetails_dbSecurityGroups :: Lens' AwsRdsDbInstanceDetails (Maybe [Text])

A list of the DB security groups to assign to the DB instance.

awsRdsDbInstanceDetails_engineVersion :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

Indicates the database engine version.

awsRdsDbInstanceDetails_deletionProtection :: Lens' AwsRdsDbInstanceDetails (Maybe Bool)

Indicates whether the DB instance has deletion protection enabled.

When deletion protection is enabled, the database cannot be deleted.

awsRdsDbInstanceDetails_storageEncrypted :: Lens' AwsRdsDbInstanceDetails (Maybe Bool)

Specifies whether the DB instance is encrypted.

awsRdsDbInstanceDetails_dbClusterIdentifier :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

awsRdsDbInstanceDetails_publiclyAccessible :: Lens' AwsRdsDbInstanceDetails (Maybe Bool)

Specifies the accessibility options for the DB instance.

A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address.

A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

awsRdsDbInstanceDetails_autoMinorVersionUpgrade :: Lens' AwsRdsDbInstanceDetails (Maybe Bool)

Indicates whether minor version patches are applied automatically.

awsRdsDbInstanceDetails_masterUsername :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The master user name of the DB instance.

awsRdsDbInstanceDetails_readReplicaDBInstanceIdentifiers :: Lens' AwsRdsDbInstanceDetails (Maybe [Text])

List of identifiers of the read replicas associated with this DB instance.

awsRdsDbInstanceDetails_iAMDatabaseAuthenticationEnabled :: Lens' AwsRdsDbInstanceDetails (Maybe Bool)

True if mapping of IAM accounts to database accounts is enabled, and otherwise false.

IAM database authentication can be enabled for the following database engines.

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher

awsRdsDbInstanceDetails_monitoringRoleArn :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The ARN for the IAM role that permits Amazon RDS to send enhanced monitoring metrics to CloudWatch Logs.

awsRdsDbInstanceDetails_iops :: Lens' AwsRdsDbInstanceDetails (Maybe Int)

Specifies the provisioned IOPS (I/O operations per second) for this DB instance.

awsRdsDbInstanceDetails_instanceCreateTime :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

Indicates when the DB instance was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsRdsDbInstanceDetails_readReplicaSourceDBInstanceIdentifier :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

If this DB instance is a read replica, contains the identifier of the source DB instance.

awsRdsDbInstanceDetails_monitoringInterval :: Lens' AwsRdsDbInstanceDetails (Maybe Int)

The interval, in seconds, between points when enhanced monitoring metrics are collected for the DB instance.

awsRdsDbInstanceDetails_engine :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

Provides the name of the database engine to use for this DB instance.

awsRdsDbInstanceDetails_processorFeatures :: Lens' AwsRdsDbInstanceDetails (Maybe [AwsRdsDbProcessorFeature])

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

awsRdsDbInstanceDetails_latestRestorableTime :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

Specifies the latest time to which a database can be restored with point-in-time restore.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsRdsDbInstanceDetails_dbInstanceClass :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

Contains the name of the compute and memory capacity class of the DB instance.

awsRdsDbInstanceDetails_promotionTier :: Lens' AwsRdsDbInstanceDetails (Maybe Int)

The order in which to promote an Aurora replica to the primary instance after a failure of the existing primary instance.

awsRdsDbInstanceDetails_licenseModel :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

License model information for this DB instance.

awsRdsDbInstanceDetails_preferredMaintenanceWindow :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Uses the format <day>:HH:MM-<day>:HH:MM.

For the day values, use mon|tue|wed|thu|fri|sat|sun.

For example, sun:09:32-sun:10:02.

awsRdsDbInstanceDetails_performanceInsightsRetentionPeriod :: Lens' AwsRdsDbInstanceDetails (Maybe Int)

The number of days to retain Performance Insights data.

awsRdsDbInstanceDetails_cACertificateIdentifier :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The identifier of the CA certificate for this DB instance.

awsRdsDbInstanceDetails_dbInstanceIdentifier :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

awsRdsDbInstanceDetails_characterSetName :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The name of the character set that this DB instance is associated with.

awsRdsDbInstanceDetails_maxAllocatedStorage :: Lens' AwsRdsDbInstanceDetails (Maybe Int)

The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

awsRdsDbInstanceDetails_kmsKeyId :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

If StorageEncrypted is true, the KMS key identifier for the encrypted DB instance.

awsRdsDbInstanceDetails_preferredBackupWindow :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The range of time each day when automated backups are created, if automated backups are enabled.

Uses the format HH:MM-HH:MM. For example, 04:52-05:22.

awsRdsDbInstanceDetails_associatedRoles :: Lens' AwsRdsDbInstanceDetails (Maybe [AwsRdsDbInstanceAssociatedRole])

The IAM roles associated with the DB instance.

awsRdsDbInstanceDetails_availabilityZone :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The Availability Zone where the DB instance will be created.

awsRdsDbInstanceDetails_vpcSecurityGroups :: Lens' AwsRdsDbInstanceDetails (Maybe [AwsRdsDbInstanceVpcSecurityGroup])

A list of VPC security groups that the DB instance belongs to.

awsRdsDbInstanceDetails_backupRetentionPeriod :: Lens' AwsRdsDbInstanceDetails (Maybe Int)

The number of days for which to retain automated backups.

awsRdsDbInstanceDetails_performanceInsightsKmsKeyId :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The identifier of the KMS key used to encrypt the Performance Insights data.

awsRdsDbInstanceDetails_dbSubnetGroup :: Lens' AwsRdsDbInstanceDetails (Maybe AwsRdsDbSubnetGroup)

Information about the subnet group that is associated with the DB instance.

awsRdsDbInstanceDetails_multiAz :: Lens' AwsRdsDbInstanceDetails (Maybe Bool)

Whether the DB instance is a multiple Availability Zone deployment.

awsRdsDbInstanceDetails_listenerEndpoint :: Lens' AwsRdsDbInstanceDetails (Maybe AwsRdsDbInstanceEndpoint)

Undocumented member.

awsRdsDbInstanceDetails_optionGroupMemberships :: Lens' AwsRdsDbInstanceDetails (Maybe [AwsRdsDbOptionGroupMembership])

The list of option group memberships for this DB instance.

awsRdsDbInstanceDetails_enhancedMonitoringResourceArn :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The ARN of the CloudWatch Logs log stream that receives the enhanced monitoring metrics data for the DB instance.

awsRdsDbInstanceDetails_secondaryAvailabilityZone :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

For a DB instance with multi-Availability Zone support, the name of the secondary Availability Zone.

awsRdsDbInstanceDetails_enabledCloudWatchLogsExports :: Lens' AwsRdsDbInstanceDetails (Maybe [Text])

A list of log types that this DB instance is configured to export to CloudWatch Logs.

awsRdsDbInstanceDetails_performanceInsightsEnabled :: Lens' AwsRdsDbInstanceDetails (Maybe Bool)

Indicates whether Performance Insights is enabled for the DB instance.

awsRdsDbInstanceDetails_allocatedStorage :: Lens' AwsRdsDbInstanceDetails (Maybe Int)

The amount of storage (in gigabytes) to initially allocate for the DB instance.

awsRdsDbInstanceDetails_dbiResourceId :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The Amazon Web Services Region-unique, immutable identifier for the DB instance. This identifier is found in CloudTrail log entries whenever the KMS key for the DB instance is accessed.

awsRdsDbInstanceDetails_dbInstanceStatus :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The current status of the DB instance.

awsRdsDbInstanceDetails_copyTagsToSnapshot :: Lens' AwsRdsDbInstanceDetails (Maybe Bool)

Whether to copy resource tags to snapshots of the DB instance.

awsRdsDbInstanceDetails_timezone :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The time zone of the DB instance.

awsRdsDbInstanceDetails_tdeCredentialArn :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The ARN from the key store with which the instance is associated for TDE encryption.

awsRdsDbInstanceDetails_dbParameterGroups :: Lens' AwsRdsDbInstanceDetails (Maybe [AwsRdsDbParameterGroup])

A list of the DB parameter groups to assign to the DB instance.

awsRdsDbInstanceDetails_endpoint :: Lens' AwsRdsDbInstanceDetails (Maybe AwsRdsDbInstanceEndpoint)

Specifies the connection endpoint.

awsRdsDbInstanceDetails_dbInstancePort :: Lens' AwsRdsDbInstanceDetails (Maybe Int)

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

awsRdsDbInstanceDetails_pendingModifiedValues :: Lens' AwsRdsDbInstanceDetails (Maybe AwsRdsDbPendingModifiedValues)

Changes to the DB instance that are currently pending.

awsRdsDbInstanceDetails_readReplicaDBClusterIdentifiers :: Lens' AwsRdsDbInstanceDetails (Maybe [Text])

List of identifiers of Aurora DB clusters to which the RDS DB instance is replicated as a read replica.

awsRdsDbInstanceDetails_storageType :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The storage type for the DB instance.

awsRdsDbInstanceDetails_statusInfos :: Lens' AwsRdsDbInstanceDetails (Maybe [AwsRdsDbStatusInfo])

The status of a read replica. If the instance isn't a read replica, this is empty.

awsRdsDbInstanceDetails_domainMemberships :: Lens' AwsRdsDbInstanceDetails (Maybe [AwsRdsDbDomainMembership])

The Active Directory domain membership records associated with the DB instance.

awsRdsDbInstanceDetails_dbName :: Lens' AwsRdsDbInstanceDetails (Maybe Text)

The meaning of this parameter differs according to the database engine you use.

MySQL, MariaDB, SQL Server, PostgreSQL

Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance.

Oracle

Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.