Amazonka.SecurityHub.Types.AwsRdsDbPendingModifiedValues

Description

Documentation

data AwsRdsDbPendingModifiedValues

Changes to a DB instance that are currently pending.

See: newAwsRdsDbPendingModifiedValues smart constructor.

Constructors

AwsRdsDbPendingModifiedValues'

Fields

engineVersion :: Maybe Text

The new engine version for the DB instance.

masterUserPassword :: Maybe Text

The new master user password for the DB instance.

dbSubnetGroupName :: Maybe Text

The name of the new subnet group for the DB instance.

iops :: Maybe Int

The new provisioned IOPS value for the DB instance.

dbInstanceClass :: Maybe Text

The new DB instance class for the DB instance.

processorFeatures :: Maybe [AwsRdsDbProcessorFeature]

Processor features that are being updated.

licenseModel :: Maybe Text

The new license model value for the DB instance.

dbInstanceIdentifier :: Maybe Text

The new DB instance identifier for the DB instance.

caCertificateIdentifier :: Maybe Text

The new CA certificate identifier for the DB instance.

pendingCloudWatchLogsExports :: Maybe AwsRdsPendingCloudWatchLogsExports

A list of log types that are being enabled or disabled.

backupRetentionPeriod :: Maybe Int

The new backup retention period for the DB instance.

multiAZ :: Maybe Bool

Indicates that a single Availability Zone DB instance is changing to a multiple Availability Zone deployment.

allocatedStorage :: Maybe Int

The new value of the allocated storage for the DB instance.

port :: Maybe Int

The new port for the DB instance.

storageType :: Maybe Text

The new storage type for the DB instance.

newAwsRdsDbPendingModifiedValues :: AwsRdsDbPendingModifiedValues

Create a value of AwsRdsDbPendingModifiedValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_engineVersion - The new engine version for the DB instance.

$sel:masterUserPassword:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_masterUserPassword - The new master user password for the DB instance.

$sel:dbSubnetGroupName:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_dbSubnetGroupName - The name of the new subnet group for the DB instance.

$sel:iops:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_iops - The new provisioned IOPS value for the DB instance.

$sel:dbInstanceClass:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_dbInstanceClass - The new DB instance class for the DB instance.

$sel:processorFeatures:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_processorFeatures - Processor features that are being updated.

$sel:licenseModel:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_licenseModel - The new license model value for the DB instance.

$sel:dbInstanceIdentifier:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_dbInstanceIdentifier - The new DB instance identifier for the DB instance.

$sel:caCertificateIdentifier:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_caCertificateIdentifier - The new CA certificate identifier for the DB instance.

$sel:pendingCloudWatchLogsExports:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_pendingCloudWatchLogsExports - A list of log types that are being enabled or disabled.

$sel:backupRetentionPeriod:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_backupRetentionPeriod - The new backup retention period for the DB instance.

$sel:multiAZ:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_multiAZ - Indicates that a single Availability Zone DB instance is changing to a multiple Availability Zone deployment.

$sel:allocatedStorage:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_allocatedStorage - The new value of the allocated storage for the DB instance.

$sel:port:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_port - The new port for the DB instance.

$sel:storageType:AwsRdsDbPendingModifiedValues', awsRdsDbPendingModifiedValues_storageType - The new storage type for the DB instance.

awsRdsDbPendingModifiedValues_engineVersion :: Lens' AwsRdsDbPendingModifiedValues (Maybe Text)

The new engine version for the DB instance.

awsRdsDbPendingModifiedValues_masterUserPassword :: Lens' AwsRdsDbPendingModifiedValues (Maybe Text)

The new master user password for the DB instance.

awsRdsDbPendingModifiedValues_dbSubnetGroupName :: Lens' AwsRdsDbPendingModifiedValues (Maybe Text)

The name of the new subnet group for the DB instance.

awsRdsDbPendingModifiedValues_iops :: Lens' AwsRdsDbPendingModifiedValues (Maybe Int)

The new provisioned IOPS value for the DB instance.

awsRdsDbPendingModifiedValues_dbInstanceClass :: Lens' AwsRdsDbPendingModifiedValues (Maybe Text)

The new DB instance class for the DB instance.

awsRdsDbPendingModifiedValues_processorFeatures :: Lens' AwsRdsDbPendingModifiedValues (Maybe [AwsRdsDbProcessorFeature])

Processor features that are being updated.

awsRdsDbPendingModifiedValues_licenseModel :: Lens' AwsRdsDbPendingModifiedValues (Maybe Text)

The new license model value for the DB instance.

awsRdsDbPendingModifiedValues_dbInstanceIdentifier :: Lens' AwsRdsDbPendingModifiedValues (Maybe Text)

The new DB instance identifier for the DB instance.

awsRdsDbPendingModifiedValues_caCertificateIdentifier :: Lens' AwsRdsDbPendingModifiedValues (Maybe Text)

The new CA certificate identifier for the DB instance.

awsRdsDbPendingModifiedValues_pendingCloudWatchLogsExports :: Lens' AwsRdsDbPendingModifiedValues (Maybe AwsRdsPendingCloudWatchLogsExports)

A list of log types that are being enabled or disabled.

awsRdsDbPendingModifiedValues_backupRetentionPeriod :: Lens' AwsRdsDbPendingModifiedValues (Maybe Int)

The new backup retention period for the DB instance.

awsRdsDbPendingModifiedValues_multiAZ :: Lens' AwsRdsDbPendingModifiedValues (Maybe Bool)

Indicates that a single Availability Zone DB instance is changing to a multiple Availability Zone deployment.

awsRdsDbPendingModifiedValues_allocatedStorage :: Lens' AwsRdsDbPendingModifiedValues (Maybe Int)

The new value of the allocated storage for the DB instance.

awsRdsDbPendingModifiedValues_port :: Lens' AwsRdsDbPendingModifiedValues (Maybe Int)

The new port for the DB instance.

awsRdsDbPendingModifiedValues_storageType :: Lens' AwsRdsDbPendingModifiedValues (Maybe Text)

The new storage type for the DB instance.