Amazonka.SecurityHub.Types.AwsRdsDbSnapshotDetails

Description

Documentation

data AwsRdsDbSnapshotDetails

Provides details about an Amazon RDS DB cluster snapshot.

See: newAwsRdsDbSnapshotDetails smart constructor.

Constructors

AwsRdsDbSnapshotDetails'

Fields

engineVersion :: Maybe Text

The version of the database engine.

status :: Maybe Text

The status of this DB snapshot.

masterUsername :: Maybe Text

The master user name for the DB snapshot.

sourceRegion :: Maybe Text

The Amazon Web Services Region that the DB snapshot was created in or copied from.

iops :: Maybe Int

The provisioned IOPS (I/O operations per second) value of the DB instance at the time of the snapshot.

iamDatabaseAuthenticationEnabled :: Maybe Bool

Whether mapping of IAM accounts to database accounts is enabled.

vpcId :: Maybe Text

The VPC ID associated with the DB snapshot.

instanceCreateTime :: Maybe Text

Specifies the time in Coordinated Universal Time (UTC) when the DB instance, from which the snapshot was taken, was created.

engine :: Maybe Text

The name of the database engine to use for this DB instance.

encrypted :: Maybe Bool

Whether the DB snapshot is encrypted.

dbSnapshotIdentifier :: Maybe Text

The name or ARN of the DB snapshot that is used to restore the DB instance.

processorFeatures :: Maybe [AwsRdsDbProcessorFeature]

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

licenseModel :: Maybe Text

License model information for the restored DB instance.

dbInstanceIdentifier :: Maybe Text

A name for the DB instance.

snapshotType :: Maybe Text

The type of the DB snapshot.

sourceDbSnapshotIdentifier :: Maybe Text

The DB snapshot ARN that the DB snapshot was copied from.

kmsKeyId :: Maybe Text

If Encrypted is true, the KMS key identifier for the encrypted DB snapshot.

availabilityZone :: Maybe Text

Specifies the name of the Availability Zone in which the DB instance was located at the time of the DB snapshot.

snapshotCreateTime :: Maybe Text

When the snapshot was taken in Coordinated Universal Time (UTC).

allocatedStorage :: Maybe Int

The amount of storage (in gigabytes) to be initially allocated for the database instance.

dbiResourceId :: Maybe Text

The identifier for the source DB instance.

optionGroupName :: Maybe Text

The option group name for the DB snapshot.

timezone :: Maybe Text

The time zone of the DB snapshot.

tdeCredentialArn :: Maybe Text

The ARN from the key store with which to associate the instance for TDE encryption.

percentProgress :: Maybe Int

The percentage of the estimated data that has been transferred.

port :: Maybe Int

The port that the database engine was listening on at the time of the snapshot.

storageType :: Maybe Text

The storage type associated with the DB snapshot.

newAwsRdsDbSnapshotDetails :: AwsRdsDbSnapshotDetails

Create a value of AwsRdsDbSnapshotDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_engineVersion - The version of the database engine.

$sel:status:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_status - The status of this DB snapshot.

$sel:masterUsername:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_masterUsername - The master user name for the DB snapshot.

$sel:sourceRegion:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_sourceRegion - The Amazon Web Services Region that the DB snapshot was created in or copied from.

$sel:iops:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_iops - The provisioned IOPS (I/O operations per second) value of the DB instance at the time of the snapshot.

$sel:iamDatabaseAuthenticationEnabled:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_iamDatabaseAuthenticationEnabled - Whether mapping of IAM accounts to database accounts is enabled.

$sel:vpcId:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_vpcId - The VPC ID associated with the DB snapshot.

$sel:instanceCreateTime:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_instanceCreateTime - Specifies the time in Coordinated Universal Time (UTC) when the DB instance, from which the snapshot was taken, was created.

$sel:engine:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_engine - The name of the database engine to use for this DB instance.

$sel:encrypted:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_encrypted - Whether the DB snapshot is encrypted.

$sel:dbSnapshotIdentifier:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_dbSnapshotIdentifier - The name or ARN of the DB snapshot that is used to restore the DB instance.

$sel:processorFeatures:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_processorFeatures - The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

$sel:licenseModel:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_licenseModel - License model information for the restored DB instance.

$sel:dbInstanceIdentifier:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_dbInstanceIdentifier - A name for the DB instance.

$sel:snapshotType:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_snapshotType - The type of the DB snapshot.

$sel:sourceDbSnapshotIdentifier:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_sourceDbSnapshotIdentifier - The DB snapshot ARN that the DB snapshot was copied from.

$sel:kmsKeyId:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_kmsKeyId - If Encrypted is true, the KMS key identifier for the encrypted DB snapshot.

$sel:availabilityZone:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_availabilityZone - Specifies the name of the Availability Zone in which the DB instance was located at the time of the DB snapshot.

$sel:snapshotCreateTime:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_snapshotCreateTime - When the snapshot was taken in Coordinated Universal Time (UTC).

$sel:allocatedStorage:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_allocatedStorage - The amount of storage (in gigabytes) to be initially allocated for the database instance.

$sel:dbiResourceId:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_dbiResourceId - The identifier for the source DB instance.

$sel:optionGroupName:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_optionGroupName - The option group name for the DB snapshot.

$sel:timezone:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_timezone - The time zone of the DB snapshot.

$sel:tdeCredentialArn:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_tdeCredentialArn - The ARN from the key store with which to associate the instance for TDE encryption.

$sel:percentProgress:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_percentProgress - The percentage of the estimated data that has been transferred.

$sel:port:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_port - The port that the database engine was listening on at the time of the snapshot.

$sel:storageType:AwsRdsDbSnapshotDetails', awsRdsDbSnapshotDetails_storageType - The storage type associated with the DB snapshot.

awsRdsDbSnapshotDetails_engineVersion :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The version of the database engine.

awsRdsDbSnapshotDetails_status :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The status of this DB snapshot.

awsRdsDbSnapshotDetails_masterUsername :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The master user name for the DB snapshot.

awsRdsDbSnapshotDetails_sourceRegion :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The Amazon Web Services Region that the DB snapshot was created in or copied from.

awsRdsDbSnapshotDetails_iops :: Lens' AwsRdsDbSnapshotDetails (Maybe Int)

The provisioned IOPS (I/O operations per second) value of the DB instance at the time of the snapshot.

awsRdsDbSnapshotDetails_iamDatabaseAuthenticationEnabled :: Lens' AwsRdsDbSnapshotDetails (Maybe Bool)

Whether mapping of IAM accounts to database accounts is enabled.

awsRdsDbSnapshotDetails_vpcId :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The VPC ID associated with the DB snapshot.

awsRdsDbSnapshotDetails_instanceCreateTime :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

Specifies the time in Coordinated Universal Time (UTC) when the DB instance, from which the snapshot was taken, was created.

awsRdsDbSnapshotDetails_engine :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The name of the database engine to use for this DB instance.

awsRdsDbSnapshotDetails_encrypted :: Lens' AwsRdsDbSnapshotDetails (Maybe Bool)

Whether the DB snapshot is encrypted.

awsRdsDbSnapshotDetails_dbSnapshotIdentifier :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The name or ARN of the DB snapshot that is used to restore the DB instance.

awsRdsDbSnapshotDetails_processorFeatures :: Lens' AwsRdsDbSnapshotDetails (Maybe [AwsRdsDbProcessorFeature])

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

awsRdsDbSnapshotDetails_licenseModel :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

License model information for the restored DB instance.

awsRdsDbSnapshotDetails_dbInstanceIdentifier :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

A name for the DB instance.

awsRdsDbSnapshotDetails_snapshotType :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The type of the DB snapshot.

awsRdsDbSnapshotDetails_sourceDbSnapshotIdentifier :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The DB snapshot ARN that the DB snapshot was copied from.

awsRdsDbSnapshotDetails_kmsKeyId :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

If Encrypted is true, the KMS key identifier for the encrypted DB snapshot.

awsRdsDbSnapshotDetails_availabilityZone :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

Specifies the name of the Availability Zone in which the DB instance was located at the time of the DB snapshot.

awsRdsDbSnapshotDetails_snapshotCreateTime :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

When the snapshot was taken in Coordinated Universal Time (UTC).

awsRdsDbSnapshotDetails_allocatedStorage :: Lens' AwsRdsDbSnapshotDetails (Maybe Int)

The amount of storage (in gigabytes) to be initially allocated for the database instance.

awsRdsDbSnapshotDetails_dbiResourceId :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The identifier for the source DB instance.

awsRdsDbSnapshotDetails_optionGroupName :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The option group name for the DB snapshot.

awsRdsDbSnapshotDetails_timezone :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The time zone of the DB snapshot.

awsRdsDbSnapshotDetails_tdeCredentialArn :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The ARN from the key store with which to associate the instance for TDE encryption.

awsRdsDbSnapshotDetails_percentProgress :: Lens' AwsRdsDbSnapshotDetails (Maybe Int)

The percentage of the estimated data that has been transferred.

awsRdsDbSnapshotDetails_port :: Lens' AwsRdsDbSnapshotDetails (Maybe Int)

The port that the database engine was listening on at the time of the snapshot.

awsRdsDbSnapshotDetails_storageType :: Lens' AwsRdsDbSnapshotDetails (Maybe Text)

The storage type associated with the DB snapshot.