Amazonka.SecurityHub.Types.AwsRdsDbSubnetGroup

Description

Documentation

data AwsRdsDbSubnetGroup

Information about the subnet group for the database instance.

See: newAwsRdsDbSubnetGroup smart constructor.

Constructors

AwsRdsDbSubnetGroup'

Fields

dbSubnetGroupName :: Maybe Text

The name of the subnet group.

vpcId :: Maybe Text

The VPC ID of the subnet group.

subnets :: Maybe [AwsRdsDbSubnetGroupSubnet]

A list of subnets in the subnet group.

dbSubnetGroupDescription :: Maybe Text

The description of the subnet group.

dbSubnetGroupArn :: Maybe Text

The ARN of the subnet group.

subnetGroupStatus :: Maybe Text

The status of the subnet group.

newAwsRdsDbSubnetGroup :: AwsRdsDbSubnetGroup

Create a value of AwsRdsDbSubnetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSubnetGroupName:AwsRdsDbSubnetGroup', awsRdsDbSubnetGroup_dbSubnetGroupName - The name of the subnet group.

$sel:vpcId:AwsRdsDbSubnetGroup', awsRdsDbSubnetGroup_vpcId - The VPC ID of the subnet group.

$sel:subnets:AwsRdsDbSubnetGroup', awsRdsDbSubnetGroup_subnets - A list of subnets in the subnet group.

$sel:dbSubnetGroupDescription:AwsRdsDbSubnetGroup', awsRdsDbSubnetGroup_dbSubnetGroupDescription - The description of the subnet group.

$sel:dbSubnetGroupArn:AwsRdsDbSubnetGroup', awsRdsDbSubnetGroup_dbSubnetGroupArn - The ARN of the subnet group.

$sel:subnetGroupStatus:AwsRdsDbSubnetGroup', awsRdsDbSubnetGroup_subnetGroupStatus - The status of the subnet group.

awsRdsDbSubnetGroup_dbSubnetGroupName :: Lens' AwsRdsDbSubnetGroup (Maybe Text)

The name of the subnet group.

awsRdsDbSubnetGroup_vpcId :: Lens' AwsRdsDbSubnetGroup (Maybe Text)

The VPC ID of the subnet group.

awsRdsDbSubnetGroup_subnets :: Lens' AwsRdsDbSubnetGroup (Maybe [AwsRdsDbSubnetGroupSubnet])

A list of subnets in the subnet group.

awsRdsDbSubnetGroup_dbSubnetGroupDescription :: Lens' AwsRdsDbSubnetGroup (Maybe Text)

The description of the subnet group.

awsRdsDbSubnetGroup_dbSubnetGroupArn :: Lens' AwsRdsDbSubnetGroup (Maybe Text)

The ARN of the subnet group.

awsRdsDbSubnetGroup_subnetGroupStatus :: Lens' AwsRdsDbSubnetGroup (Maybe Text)

The status of the subnet group.