Amazonka.SecurityHub.Types.AwsRedshiftClusterClusterSnapshotCopyStatus

Description

Documentation

data AwsRedshiftClusterClusterSnapshotCopyStatus

Information about a cross-Region snapshot copy.

See: newAwsRedshiftClusterClusterSnapshotCopyStatus smart constructor.

Constructors

AwsRedshiftClusterClusterSnapshotCopyStatus'

Fields

manualSnapshotRetentionPeriod :: Maybe Int

The number of days that manual snapshots are retained in the destination region after they are copied from a source region.

If the value is -1, then the manual snapshot is retained indefinitely.

Valid values: Either -1 or an integer between 1 and 3,653

retentionPeriod :: Maybe Int

The number of days to retain automated snapshots in the destination Region after they are copied from a source Region.

destinationRegion :: Maybe Text

The destination Region that snapshots are automatically copied to when cross-Region snapshot copy is enabled.

snapshotCopyGrantName :: Maybe Text

The name of the snapshot copy grant.

newAwsRedshiftClusterClusterSnapshotCopyStatus :: AwsRedshiftClusterClusterSnapshotCopyStatus

Create a value of AwsRedshiftClusterClusterSnapshotCopyStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manualSnapshotRetentionPeriod:AwsRedshiftClusterClusterSnapshotCopyStatus', awsRedshiftClusterClusterSnapshotCopyStatus_manualSnapshotRetentionPeriod - The number of days that manual snapshots are retained in the destination region after they are copied from a source region.

If the value is -1, then the manual snapshot is retained indefinitely.

Valid values: Either -1 or an integer between 1 and 3,653

$sel:retentionPeriod:AwsRedshiftClusterClusterSnapshotCopyStatus', awsRedshiftClusterClusterSnapshotCopyStatus_retentionPeriod - The number of days to retain automated snapshots in the destination Region after they are copied from a source Region.

$sel:destinationRegion:AwsRedshiftClusterClusterSnapshotCopyStatus', awsRedshiftClusterClusterSnapshotCopyStatus_destinationRegion - The destination Region that snapshots are automatically copied to when cross-Region snapshot copy is enabled.

$sel:snapshotCopyGrantName:AwsRedshiftClusterClusterSnapshotCopyStatus', awsRedshiftClusterClusterSnapshotCopyStatus_snapshotCopyGrantName - The name of the snapshot copy grant.

awsRedshiftClusterClusterSnapshotCopyStatus_manualSnapshotRetentionPeriod :: Lens' AwsRedshiftClusterClusterSnapshotCopyStatus (Maybe Int)

The number of days that manual snapshots are retained in the destination region after they are copied from a source region.

If the value is -1, then the manual snapshot is retained indefinitely.

Valid values: Either -1 or an integer between 1 and 3,653

awsRedshiftClusterClusterSnapshotCopyStatus_retentionPeriod :: Lens' AwsRedshiftClusterClusterSnapshotCopyStatus (Maybe Int)

The number of days to retain automated snapshots in the destination Region after they are copied from a source Region.

awsRedshiftClusterClusterSnapshotCopyStatus_destinationRegion :: Lens' AwsRedshiftClusterClusterSnapshotCopyStatus (Maybe Text)

The destination Region that snapshots are automatically copied to when cross-Region snapshot copy is enabled.

awsRedshiftClusterClusterSnapshotCopyStatus_snapshotCopyGrantName :: Lens' AwsRedshiftClusterClusterSnapshotCopyStatus (Maybe Text)

The name of the snapshot copy grant.