Amazonka.SecurityHub.Types.AwsRedshiftClusterDetails

Description

Documentation

data AwsRedshiftClusterDetails

Details about an Amazon Redshift cluster.

See: newAwsRedshiftClusterDetails smart constructor.

Constructors

AwsRedshiftClusterDetails'

Fields

resizeInfo :: Maybe AwsRedshiftClusterResizeInfo

Information about the resize operation for the cluster.

restoreStatus :: Maybe AwsRedshiftClusterRestoreStatus

Information about the status of a cluster restore action. Only applies to a cluster that was created by restoring a snapshot.

manualSnapshotRetentionPeriod :: Maybe Int

The default number of days to retain a manual snapshot.

If the value is -1, the snapshot is retained indefinitely.

This setting doesn't change the retention period of existing snapshots.

Valid values: Either -1 or an integer between 1 and 3,653

enhancedVpcRouting :: Maybe Bool

Indicates whether to create the cluster with enhanced VPC routing enabled.

clusterSnapshotCopyStatus :: Maybe AwsRedshiftClusterClusterSnapshotCopyStatus

Information about the destination Region and retention period for the cross-Region snapshot copy.

clusterAvailabilityStatus :: Maybe Text

The availability status of the cluster for queries. Possible values are the following:

Available - The cluster is available for queries.

Unavailable - The cluster is not available for queries.

Maintenance - The cluster is intermittently available for queries due to maintenance activities.

Modifying -The cluster is intermittently available for queries due to changes that modify the cluster.

Failed - The cluster failed and is not available for queries.

clusterRevisionNumber :: Maybe Text

The specific revision number of the database in the cluster.

snapshotScheduleIdentifier :: Maybe Text

A unique identifier for the cluster snapshot schedule.

publiclyAccessible :: Maybe Bool

Whether the cluster can be accessed from a public network.

masterUsername :: Maybe Text

The master user name for the cluster. This name is used to connect to the database that is specified in as the value of DBName.

maintenanceTrackName :: Maybe Text

The name of the maintenance track for the cluster.

expectedNextSnapshotScheduleTime :: Maybe Text

Indicates when the next snapshot is expected to be taken. The cluster must have a valid snapshot schedule and have backups enabled.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

elasticResizeNumberOfNodeOptions :: Maybe Text

The number of nodes that you can use the elastic resize method to resize the cluster to.

vpcId :: Maybe Text

The identifier of the VPC that the cluster is in, if the cluster is in a VPC.

clusterSecurityGroups :: Maybe [AwsRedshiftClusterClusterSecurityGroup]

A list of cluster security groups that are associated with the cluster.

automatedSnapshotRetentionPeriod :: Maybe Int

The number of days that automatic cluster snapshots are retained.

snapshotScheduleState :: Maybe Text

The current state of the cluster snapshot schedule.

Valid values: MODIFYING | ACTIVE | FAILED

encrypted :: Maybe Bool

Indicates whether the data in the cluster is encrypted at rest.

clusterSubnetGroupName :: Maybe Text

The name of the subnet group that is associated with the cluster. This parameter is valid only when the cluster is in a VPC.

expectedNextSnapshotScheduleTimeStatus :: Maybe Text

The status of the next expected snapshot.

Valid values: OnTrack | Pending

clusterIdentifier :: Maybe Text

The unique identifier of the cluster.

deferredMaintenanceWindows :: Maybe [AwsRedshiftClusterDeferredMaintenanceWindow]

List of time windows during which maintenance was deferred.

numberOfNodes :: Maybe Int

The number of compute nodes in the cluster.

clusterPublicKey :: Maybe Text

The public key for the cluster.

preferredMaintenanceWindow :: Maybe Text

The weekly time range, in Universal Coordinated Time (UTC), during which system maintenance can occur.

Format: <day>:HH:MM-<day>:HH:MM

For the day values, use mon | tue | wed | thu | fri | sat | sun

For example, sun:09:32-sun:10:02

kmsKeyId :: Maybe Text

The identifier of the KMS encryption key that is used to encrypt data in the cluster.

clusterParameterGroups :: Maybe [AwsRedshiftClusterClusterParameterGroup]

The list of cluster parameter groups that are associated with this cluster.

availabilityZone :: Maybe Text

The name of the Availability Zone in which the cluster is located.

vpcSecurityGroups :: Maybe [AwsRedshiftClusterVpcSecurityGroup]

The list of VPC security groups that the cluster belongs to, if the cluster is in a VPC.

hsmStatus :: Maybe AwsRedshiftClusterHsmStatus

Information about whether the Amazon Redshift cluster finished applying any changes to hardware security module (HSM) settings that were specified in a modify cluster command.

iamRoles :: Maybe [AwsRedshiftClusterIamRole]

A list of IAM roles that the cluster can use to access other Amazon Web Services services.

pendingActions :: Maybe [Text]

A list of cluster operations that are waiting to start.

elasticIpStatus :: Maybe AwsRedshiftClusterElasticIpStatus

Information about the status of the Elastic IP (EIP) address.

clusterVersion :: Maybe Text

The version ID of the Amazon Redshift engine that runs on the cluster.

nodeType :: Maybe Text

The node type for the nodes in the cluster.

nextMaintenanceWindowStartTime :: Maybe Text

Indicates the start of the next maintenance window.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

clusterCreateTime :: Maybe Text

Indicates when the cluster was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

endpoint :: Maybe AwsRedshiftClusterEndpoint

The connection endpoint.

allowVersionUpgrade :: Maybe Bool

Indicates whether major version upgrades are applied automatically to the cluster during the maintenance window.

clusterStatus :: Maybe Text

The current status of the cluster.

Valid values: available | available, prep-for-resize | available, resize-cleanup | cancelling-resize | creating | deleting | final-snapshot | hardware-failure | incompatible-hsm | incompatible-network | incompatible-parameters | incompatible-restore | modifying | paused | rebooting | renaming | resizing | rotating-keys | storage-full | updating-hsm

pendingModifiedValues :: Maybe AwsRedshiftClusterPendingModifiedValues

A list of changes to the cluster that are currently pending.

clusterNodes :: Maybe [AwsRedshiftClusterClusterNode]

The nodes in the cluster.

dbName :: Maybe Text

The name of the initial database that was created when the cluster was created.

The same name is returned for the life of the cluster.

If an initial database is not specified, a database named devdev is created by default.

newAwsRedshiftClusterDetails :: AwsRedshiftClusterDetails

Create a value of AwsRedshiftClusterDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resizeInfo:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_resizeInfo - Information about the resize operation for the cluster.

$sel:restoreStatus:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_restoreStatus - Information about the status of a cluster restore action. Only applies to a cluster that was created by restoring a snapshot.

$sel:manualSnapshotRetentionPeriod:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_manualSnapshotRetentionPeriod - The default number of days to retain a manual snapshot.

If the value is -1, the snapshot is retained indefinitely.

This setting doesn't change the retention period of existing snapshots.

Valid values: Either -1 or an integer between 1 and 3,653

$sel:enhancedVpcRouting:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_enhancedVpcRouting - Indicates whether to create the cluster with enhanced VPC routing enabled.

$sel:clusterSnapshotCopyStatus:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_clusterSnapshotCopyStatus - Information about the destination Region and retention period for the cross-Region snapshot copy.

$sel:clusterAvailabilityStatus:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_clusterAvailabilityStatus - The availability status of the cluster for queries. Possible values are the following:

Available - The cluster is available for queries.

Unavailable - The cluster is not available for queries.

Maintenance - The cluster is intermittently available for queries due to maintenance activities.

Modifying -The cluster is intermittently available for queries due to changes that modify the cluster.

Failed - The cluster failed and is not available for queries.

$sel:clusterRevisionNumber:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_clusterRevisionNumber - The specific revision number of the database in the cluster.

$sel:snapshotScheduleIdentifier:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_snapshotScheduleIdentifier - A unique identifier for the cluster snapshot schedule.

$sel:publiclyAccessible:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_publiclyAccessible - Whether the cluster can be accessed from a public network.

$sel:masterUsername:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_masterUsername - The master user name for the cluster. This name is used to connect to the database that is specified in as the value of DBName.

$sel:maintenanceTrackName:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_maintenanceTrackName - The name of the maintenance track for the cluster.

$sel:expectedNextSnapshotScheduleTime:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_expectedNextSnapshotScheduleTime - Indicates when the next snapshot is expected to be taken. The cluster must have a valid snapshot schedule and have backups enabled.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:elasticResizeNumberOfNodeOptions:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_elasticResizeNumberOfNodeOptions - The number of nodes that you can use the elastic resize method to resize the cluster to.

$sel:vpcId:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_vpcId - The identifier of the VPC that the cluster is in, if the cluster is in a VPC.

$sel:clusterSecurityGroups:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_clusterSecurityGroups - A list of cluster security groups that are associated with the cluster.

$sel:automatedSnapshotRetentionPeriod:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_automatedSnapshotRetentionPeriod - The number of days that automatic cluster snapshots are retained.

$sel:snapshotScheduleState:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_snapshotScheduleState - The current state of the cluster snapshot schedule.

Valid values: MODIFYING | ACTIVE | FAILED

$sel:encrypted:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_encrypted - Indicates whether the data in the cluster is encrypted at rest.

$sel:clusterSubnetGroupName:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_clusterSubnetGroupName - The name of the subnet group that is associated with the cluster. This parameter is valid only when the cluster is in a VPC.

$sel:expectedNextSnapshotScheduleTimeStatus:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_expectedNextSnapshotScheduleTimeStatus - The status of the next expected snapshot.

Valid values: OnTrack | Pending

$sel:clusterIdentifier:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_clusterIdentifier - The unique identifier of the cluster.

$sel:deferredMaintenanceWindows:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_deferredMaintenanceWindows - List of time windows during which maintenance was deferred.

$sel:numberOfNodes:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_numberOfNodes - The number of compute nodes in the cluster.

$sel:clusterPublicKey:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_clusterPublicKey - The public key for the cluster.

$sel:preferredMaintenanceWindow:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_preferredMaintenanceWindow - The weekly time range, in Universal Coordinated Time (UTC), during which system maintenance can occur.

Format: <day>:HH:MM-<day>:HH:MM

For the day values, use mon | tue | wed | thu | fri | sat | sun

For example, sun:09:32-sun:10:02

$sel:kmsKeyId:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_kmsKeyId - The identifier of the KMS encryption key that is used to encrypt data in the cluster.

$sel:clusterParameterGroups:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_clusterParameterGroups - The list of cluster parameter groups that are associated with this cluster.

$sel:availabilityZone:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_availabilityZone - The name of the Availability Zone in which the cluster is located.

$sel:vpcSecurityGroups:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_vpcSecurityGroups - The list of VPC security groups that the cluster belongs to, if the cluster is in a VPC.

$sel:hsmStatus:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_hsmStatus - Information about whether the Amazon Redshift cluster finished applying any changes to hardware security module (HSM) settings that were specified in a modify cluster command.

$sel:iamRoles:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_iamRoles - A list of IAM roles that the cluster can use to access other Amazon Web Services services.

$sel:pendingActions:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_pendingActions - A list of cluster operations that are waiting to start.

$sel:elasticIpStatus:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_elasticIpStatus - Information about the status of the Elastic IP (EIP) address.

$sel:clusterVersion:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_clusterVersion - The version ID of the Amazon Redshift engine that runs on the cluster.

$sel:nodeType:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_nodeType - The node type for the nodes in the cluster.

$sel:nextMaintenanceWindowStartTime:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_nextMaintenanceWindowStartTime - Indicates the start of the next maintenance window.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:clusterCreateTime:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_clusterCreateTime - Indicates when the cluster was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:endpoint:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_endpoint - The connection endpoint.

$sel:allowVersionUpgrade:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_allowVersionUpgrade - Indicates whether major version upgrades are applied automatically to the cluster during the maintenance window.

$sel:clusterStatus:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_clusterStatus - The current status of the cluster.

Valid values: available | available, prep-for-resize | available, resize-cleanup | cancelling-resize | creating | deleting | final-snapshot | hardware-failure | incompatible-hsm | incompatible-network | incompatible-parameters | incompatible-restore | modifying | paused | rebooting | renaming | resizing | rotating-keys | storage-full | updating-hsm

$sel:pendingModifiedValues:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_pendingModifiedValues - A list of changes to the cluster that are currently pending.

$sel:clusterNodes:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_clusterNodes - The nodes in the cluster.

$sel:dbName:AwsRedshiftClusterDetails', awsRedshiftClusterDetails_dbName - The name of the initial database that was created when the cluster was created.

The same name is returned for the life of the cluster.

If an initial database is not specified, a database named devdev is created by default.

awsRedshiftClusterDetails_resizeInfo :: Lens' AwsRedshiftClusterDetails (Maybe AwsRedshiftClusterResizeInfo)

Information about the resize operation for the cluster.

awsRedshiftClusterDetails_restoreStatus :: Lens' AwsRedshiftClusterDetails (Maybe AwsRedshiftClusterRestoreStatus)

Information about the status of a cluster restore action. Only applies to a cluster that was created by restoring a snapshot.

awsRedshiftClusterDetails_manualSnapshotRetentionPeriod :: Lens' AwsRedshiftClusterDetails (Maybe Int)

The default number of days to retain a manual snapshot.

If the value is -1, the snapshot is retained indefinitely.

This setting doesn't change the retention period of existing snapshots.

Valid values: Either -1 or an integer between 1 and 3,653

awsRedshiftClusterDetails_enhancedVpcRouting :: Lens' AwsRedshiftClusterDetails (Maybe Bool)

Indicates whether to create the cluster with enhanced VPC routing enabled.

awsRedshiftClusterDetails_clusterSnapshotCopyStatus :: Lens' AwsRedshiftClusterDetails (Maybe AwsRedshiftClusterClusterSnapshotCopyStatus)

Information about the destination Region and retention period for the cross-Region snapshot copy.

awsRedshiftClusterDetails_clusterAvailabilityStatus :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The availability status of the cluster for queries. Possible values are the following:

Available - The cluster is available for queries.

Unavailable - The cluster is not available for queries.

Maintenance - The cluster is intermittently available for queries due to maintenance activities.

Modifying -The cluster is intermittently available for queries due to changes that modify the cluster.

Failed - The cluster failed and is not available for queries.

awsRedshiftClusterDetails_clusterRevisionNumber :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The specific revision number of the database in the cluster.

awsRedshiftClusterDetails_snapshotScheduleIdentifier :: Lens' AwsRedshiftClusterDetails (Maybe Text)

A unique identifier for the cluster snapshot schedule.

awsRedshiftClusterDetails_publiclyAccessible :: Lens' AwsRedshiftClusterDetails (Maybe Bool)

Whether the cluster can be accessed from a public network.

awsRedshiftClusterDetails_masterUsername :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The master user name for the cluster. This name is used to connect to the database that is specified in as the value of DBName.

awsRedshiftClusterDetails_maintenanceTrackName :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The name of the maintenance track for the cluster.

awsRedshiftClusterDetails_expectedNextSnapshotScheduleTime :: Lens' AwsRedshiftClusterDetails (Maybe Text)

Indicates when the next snapshot is expected to be taken. The cluster must have a valid snapshot schedule and have backups enabled.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsRedshiftClusterDetails_elasticResizeNumberOfNodeOptions :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The number of nodes that you can use the elastic resize method to resize the cluster to.

awsRedshiftClusterDetails_vpcId :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The identifier of the VPC that the cluster is in, if the cluster is in a VPC.

awsRedshiftClusterDetails_clusterSecurityGroups :: Lens' AwsRedshiftClusterDetails (Maybe [AwsRedshiftClusterClusterSecurityGroup])

A list of cluster security groups that are associated with the cluster.

awsRedshiftClusterDetails_automatedSnapshotRetentionPeriod :: Lens' AwsRedshiftClusterDetails (Maybe Int)

The number of days that automatic cluster snapshots are retained.

awsRedshiftClusterDetails_snapshotScheduleState :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The current state of the cluster snapshot schedule.

Valid values: MODIFYING | ACTIVE | FAILED

awsRedshiftClusterDetails_encrypted :: Lens' AwsRedshiftClusterDetails (Maybe Bool)

Indicates whether the data in the cluster is encrypted at rest.

awsRedshiftClusterDetails_clusterSubnetGroupName :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The name of the subnet group that is associated with the cluster. This parameter is valid only when the cluster is in a VPC.

awsRedshiftClusterDetails_expectedNextSnapshotScheduleTimeStatus :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The status of the next expected snapshot.

Valid values: OnTrack | Pending

awsRedshiftClusterDetails_clusterIdentifier :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The unique identifier of the cluster.

awsRedshiftClusterDetails_deferredMaintenanceWindows :: Lens' AwsRedshiftClusterDetails (Maybe [AwsRedshiftClusterDeferredMaintenanceWindow])

List of time windows during which maintenance was deferred.

awsRedshiftClusterDetails_numberOfNodes :: Lens' AwsRedshiftClusterDetails (Maybe Int)

The number of compute nodes in the cluster.

awsRedshiftClusterDetails_clusterPublicKey :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The public key for the cluster.

awsRedshiftClusterDetails_preferredMaintenanceWindow :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The weekly time range, in Universal Coordinated Time (UTC), during which system maintenance can occur.

Format: <day>:HH:MM-<day>:HH:MM

For the day values, use mon | tue | wed | thu | fri | sat | sun

For example, sun:09:32-sun:10:02

awsRedshiftClusterDetails_kmsKeyId :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The identifier of the KMS encryption key that is used to encrypt data in the cluster.

awsRedshiftClusterDetails_clusterParameterGroups :: Lens' AwsRedshiftClusterDetails (Maybe [AwsRedshiftClusterClusterParameterGroup])

The list of cluster parameter groups that are associated with this cluster.

awsRedshiftClusterDetails_availabilityZone :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The name of the Availability Zone in which the cluster is located.

awsRedshiftClusterDetails_vpcSecurityGroups :: Lens' AwsRedshiftClusterDetails (Maybe [AwsRedshiftClusterVpcSecurityGroup])

The list of VPC security groups that the cluster belongs to, if the cluster is in a VPC.

awsRedshiftClusterDetails_hsmStatus :: Lens' AwsRedshiftClusterDetails (Maybe AwsRedshiftClusterHsmStatus)

Information about whether the Amazon Redshift cluster finished applying any changes to hardware security module (HSM) settings that were specified in a modify cluster command.

awsRedshiftClusterDetails_iamRoles :: Lens' AwsRedshiftClusterDetails (Maybe [AwsRedshiftClusterIamRole])

A list of IAM roles that the cluster can use to access other Amazon Web Services services.

awsRedshiftClusterDetails_pendingActions :: Lens' AwsRedshiftClusterDetails (Maybe [Text])

A list of cluster operations that are waiting to start.

awsRedshiftClusterDetails_elasticIpStatus :: Lens' AwsRedshiftClusterDetails (Maybe AwsRedshiftClusterElasticIpStatus)

Information about the status of the Elastic IP (EIP) address.

awsRedshiftClusterDetails_clusterVersion :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The version ID of the Amazon Redshift engine that runs on the cluster.

awsRedshiftClusterDetails_nodeType :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The node type for the nodes in the cluster.

awsRedshiftClusterDetails_nextMaintenanceWindowStartTime :: Lens' AwsRedshiftClusterDetails (Maybe Text)

Indicates the start of the next maintenance window.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsRedshiftClusterDetails_clusterCreateTime :: Lens' AwsRedshiftClusterDetails (Maybe Text)

Indicates when the cluster was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsRedshiftClusterDetails_endpoint :: Lens' AwsRedshiftClusterDetails (Maybe AwsRedshiftClusterEndpoint)

The connection endpoint.

awsRedshiftClusterDetails_allowVersionUpgrade :: Lens' AwsRedshiftClusterDetails (Maybe Bool)

Indicates whether major version upgrades are applied automatically to the cluster during the maintenance window.

awsRedshiftClusterDetails_clusterStatus :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The current status of the cluster.

Valid values: available | available, prep-for-resize | available, resize-cleanup | cancelling-resize | creating | deleting | final-snapshot | hardware-failure | incompatible-hsm | incompatible-network | incompatible-parameters | incompatible-restore | modifying | paused | rebooting | renaming | resizing | rotating-keys | storage-full | updating-hsm

awsRedshiftClusterDetails_pendingModifiedValues :: Lens' AwsRedshiftClusterDetails (Maybe AwsRedshiftClusterPendingModifiedValues)

A list of changes to the cluster that are currently pending.

awsRedshiftClusterDetails_clusterNodes :: Lens' AwsRedshiftClusterDetails (Maybe [AwsRedshiftClusterClusterNode])

The nodes in the cluster.

awsRedshiftClusterDetails_dbName :: Lens' AwsRedshiftClusterDetails (Maybe Text)

The name of the initial database that was created when the cluster was created.

The same name is returned for the life of the cluster.

If an initial database is not specified, a database named devdev is created by default.