Amazonka.SecurityHub.Types.AwsRedshiftClusterPendingModifiedValues

Description

Documentation

data AwsRedshiftClusterPendingModifiedValues

Changes to the Amazon Redshift cluster that are currently pending.

See: newAwsRedshiftClusterPendingModifiedValues smart constructor.

Constructors

AwsRedshiftClusterPendingModifiedValues'

Fields

encryptionType :: Maybe Text

The encryption type for a cluster.

enhancedVpcRouting :: Maybe Bool

Indicates whether to create the cluster with enhanced VPC routing enabled.

masterUserPassword :: Maybe Text

The pending or in-progress change to the master user password for the cluster.

publiclyAccessible :: Maybe Bool

The pending or in-progress change to whether the cluster can be connected to from the public network.

maintenanceTrackName :: Maybe Text

The name of the maintenance track that the cluster changes to during the next maintenance window.

automatedSnapshotRetentionPeriod :: Maybe Int

The pending or in-progress change to the automated snapshot retention period.

clusterIdentifier :: Maybe Text

The pending or in-progress change to the identifier for the cluster.

numberOfNodes :: Maybe Int

The pending or in-progress change to the number of nodes in the cluster.

clusterType :: Maybe Text

The pending or in-progress change to the cluster type.

clusterVersion :: Maybe Text

The pending or in-progress change to the service version.

nodeType :: Maybe Text

The pending or in-progress change to the cluster's node type.

newAwsRedshiftClusterPendingModifiedValues :: AwsRedshiftClusterPendingModifiedValues

Create a value of AwsRedshiftClusterPendingModifiedValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionType:AwsRedshiftClusterPendingModifiedValues', awsRedshiftClusterPendingModifiedValues_encryptionType - The encryption type for a cluster.

$sel:enhancedVpcRouting:AwsRedshiftClusterPendingModifiedValues', awsRedshiftClusterPendingModifiedValues_enhancedVpcRouting - Indicates whether to create the cluster with enhanced VPC routing enabled.

$sel:masterUserPassword:AwsRedshiftClusterPendingModifiedValues', awsRedshiftClusterPendingModifiedValues_masterUserPassword - The pending or in-progress change to the master user password for the cluster.

$sel:publiclyAccessible:AwsRedshiftClusterPendingModifiedValues', awsRedshiftClusterPendingModifiedValues_publiclyAccessible - The pending or in-progress change to whether the cluster can be connected to from the public network.

$sel:maintenanceTrackName:AwsRedshiftClusterPendingModifiedValues', awsRedshiftClusterPendingModifiedValues_maintenanceTrackName - The name of the maintenance track that the cluster changes to during the next maintenance window.

$sel:automatedSnapshotRetentionPeriod:AwsRedshiftClusterPendingModifiedValues', awsRedshiftClusterPendingModifiedValues_automatedSnapshotRetentionPeriod - The pending or in-progress change to the automated snapshot retention period.

$sel:clusterIdentifier:AwsRedshiftClusterPendingModifiedValues', awsRedshiftClusterPendingModifiedValues_clusterIdentifier - The pending or in-progress change to the identifier for the cluster.

$sel:numberOfNodes:AwsRedshiftClusterPendingModifiedValues', awsRedshiftClusterPendingModifiedValues_numberOfNodes - The pending or in-progress change to the number of nodes in the cluster.

$sel:clusterType:AwsRedshiftClusterPendingModifiedValues', awsRedshiftClusterPendingModifiedValues_clusterType - The pending or in-progress change to the cluster type.

$sel:clusterVersion:AwsRedshiftClusterPendingModifiedValues', awsRedshiftClusterPendingModifiedValues_clusterVersion - The pending or in-progress change to the service version.

$sel:nodeType:AwsRedshiftClusterPendingModifiedValues', awsRedshiftClusterPendingModifiedValues_nodeType - The pending or in-progress change to the cluster's node type.

awsRedshiftClusterPendingModifiedValues_encryptionType :: Lens' AwsRedshiftClusterPendingModifiedValues (Maybe Text)

The encryption type for a cluster.

awsRedshiftClusterPendingModifiedValues_enhancedVpcRouting :: Lens' AwsRedshiftClusterPendingModifiedValues (Maybe Bool)

Indicates whether to create the cluster with enhanced VPC routing enabled.

awsRedshiftClusterPendingModifiedValues_masterUserPassword :: Lens' AwsRedshiftClusterPendingModifiedValues (Maybe Text)

The pending or in-progress change to the master user password for the cluster.

awsRedshiftClusterPendingModifiedValues_publiclyAccessible :: Lens' AwsRedshiftClusterPendingModifiedValues (Maybe Bool)

The pending or in-progress change to whether the cluster can be connected to from the public network.

awsRedshiftClusterPendingModifiedValues_maintenanceTrackName :: Lens' AwsRedshiftClusterPendingModifiedValues (Maybe Text)

The name of the maintenance track that the cluster changes to during the next maintenance window.

awsRedshiftClusterPendingModifiedValues_automatedSnapshotRetentionPeriod :: Lens' AwsRedshiftClusterPendingModifiedValues (Maybe Int)

The pending or in-progress change to the automated snapshot retention period.

awsRedshiftClusterPendingModifiedValues_clusterIdentifier :: Lens' AwsRedshiftClusterPendingModifiedValues (Maybe Text)

The pending or in-progress change to the identifier for the cluster.

awsRedshiftClusterPendingModifiedValues_numberOfNodes :: Lens' AwsRedshiftClusterPendingModifiedValues (Maybe Int)

The pending or in-progress change to the number of nodes in the cluster.

awsRedshiftClusterPendingModifiedValues_clusterType :: Lens' AwsRedshiftClusterPendingModifiedValues (Maybe Text)

The pending or in-progress change to the cluster type.

awsRedshiftClusterPendingModifiedValues_clusterVersion :: Lens' AwsRedshiftClusterPendingModifiedValues (Maybe Text)

The pending or in-progress change to the service version.

awsRedshiftClusterPendingModifiedValues_nodeType :: Lens' AwsRedshiftClusterPendingModifiedValues (Maybe Text)

The pending or in-progress change to the cluster's node type.