Amazonka.SecurityHub.Types.AwsS3BucketBucketLifecycleConfigurationRulesDetails

Description

Documentation

data AwsS3BucketBucketLifecycleConfigurationRulesDetails

Configuration for a lifecycle rule.

See: newAwsS3BucketBucketLifecycleConfigurationRulesDetails smart constructor.

Constructors

AwsS3BucketBucketLifecycleConfigurationRulesDetails'

Fields

status :: Maybe Text

The current status of the rule. Indicates whether the rule is currently being applied.

transitions :: Maybe [AwsS3BucketBucketLifecycleConfigurationRulesTransitionsDetails]

Transition rules that indicate when objects transition to a specified storage class.

prefix :: Maybe Text

A prefix that identifies one or more objects that the rule applies to.

noncurrentVersionTransitions :: Maybe [AwsS3BucketBucketLifecycleConfigurationRulesNoncurrentVersionTransitionsDetails]

Transition rules that describe when noncurrent objects transition to a specified storage class.

noncurrentVersionExpirationInDays :: Maybe Int

The number of days that an object is noncurrent before Amazon S3 can perform the associated action.

expirationDate :: Maybe Text

The date when objects are moved or deleted.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

id :: Maybe Text

The unique identifier of the rule.

filter' :: Maybe AwsS3BucketBucketLifecycleConfigurationRulesFilterDetails

Identifies the objects that a rule applies to.

expirationInDays :: Maybe Int

The length in days of the lifetime for objects that are subject to the rule.

abortIncompleteMultipartUpload :: Maybe AwsS3BucketBucketLifecycleConfigurationRulesAbortIncompleteMultipartUploadDetails

How Amazon S3 responds when a multipart upload is incomplete. Specifically, provides a number of days before Amazon S3 cancels the entire upload.

expiredObjectDeleteMarker :: Maybe Bool

Whether Amazon S3 removes a delete marker that has no noncurrent versions. If set to true, the delete marker is expired. If set to false, the policy takes no action.

If you provide ExpiredObjectDeleteMarker, you cannot provide ExpirationInDays or ExpirationDate.

newAwsS3BucketBucketLifecycleConfigurationRulesDetails :: AwsS3BucketBucketLifecycleConfigurationRulesDetails

Create a value of AwsS3BucketBucketLifecycleConfigurationRulesDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AwsS3BucketBucketLifecycleConfigurationRulesDetails', awsS3BucketBucketLifecycleConfigurationRulesDetails_status - The current status of the rule. Indicates whether the rule is currently being applied.

$sel:transitions:AwsS3BucketBucketLifecycleConfigurationRulesDetails', awsS3BucketBucketLifecycleConfigurationRulesDetails_transitions - Transition rules that indicate when objects transition to a specified storage class.

$sel:prefix:AwsS3BucketBucketLifecycleConfigurationRulesDetails', awsS3BucketBucketLifecycleConfigurationRulesDetails_prefix - A prefix that identifies one or more objects that the rule applies to.

$sel:noncurrentVersionTransitions:AwsS3BucketBucketLifecycleConfigurationRulesDetails', awsS3BucketBucketLifecycleConfigurationRulesDetails_noncurrentVersionTransitions - Transition rules that describe when noncurrent objects transition to a specified storage class.

$sel:noncurrentVersionExpirationInDays:AwsS3BucketBucketLifecycleConfigurationRulesDetails', awsS3BucketBucketLifecycleConfigurationRulesDetails_noncurrentVersionExpirationInDays - The number of days that an object is noncurrent before Amazon S3 can perform the associated action.

$sel:expirationDate:AwsS3BucketBucketLifecycleConfigurationRulesDetails', awsS3BucketBucketLifecycleConfigurationRulesDetails_expirationDate - The date when objects are moved or deleted.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:id:AwsS3BucketBucketLifecycleConfigurationRulesDetails', awsS3BucketBucketLifecycleConfigurationRulesDetails_id - The unique identifier of the rule.

$sel:filter':AwsS3BucketBucketLifecycleConfigurationRulesDetails', awsS3BucketBucketLifecycleConfigurationRulesDetails_filter - Identifies the objects that a rule applies to.

$sel:expirationInDays:AwsS3BucketBucketLifecycleConfigurationRulesDetails', awsS3BucketBucketLifecycleConfigurationRulesDetails_expirationInDays - The length in days of the lifetime for objects that are subject to the rule.

$sel:abortIncompleteMultipartUpload:AwsS3BucketBucketLifecycleConfigurationRulesDetails', awsS3BucketBucketLifecycleConfigurationRulesDetails_abortIncompleteMultipartUpload - How Amazon S3 responds when a multipart upload is incomplete. Specifically, provides a number of days before Amazon S3 cancels the entire upload.

$sel:expiredObjectDeleteMarker:AwsS3BucketBucketLifecycleConfigurationRulesDetails', awsS3BucketBucketLifecycleConfigurationRulesDetails_expiredObjectDeleteMarker - Whether Amazon S3 removes a delete marker that has no noncurrent versions. If set to true, the delete marker is expired. If set to false, the policy takes no action.

If you provide ExpiredObjectDeleteMarker, you cannot provide ExpirationInDays or ExpirationDate.

awsS3BucketBucketLifecycleConfigurationRulesDetails_status :: Lens' AwsS3BucketBucketLifecycleConfigurationRulesDetails (Maybe Text)

The current status of the rule. Indicates whether the rule is currently being applied.

awsS3BucketBucketLifecycleConfigurationRulesDetails_transitions :: Lens' AwsS3BucketBucketLifecycleConfigurationRulesDetails (Maybe [AwsS3BucketBucketLifecycleConfigurationRulesTransitionsDetails])

Transition rules that indicate when objects transition to a specified storage class.

awsS3BucketBucketLifecycleConfigurationRulesDetails_prefix :: Lens' AwsS3BucketBucketLifecycleConfigurationRulesDetails (Maybe Text)

A prefix that identifies one or more objects that the rule applies to.

awsS3BucketBucketLifecycleConfigurationRulesDetails_noncurrentVersionTransitions :: Lens' AwsS3BucketBucketLifecycleConfigurationRulesDetails (Maybe [AwsS3BucketBucketLifecycleConfigurationRulesNoncurrentVersionTransitionsDetails])

Transition rules that describe when noncurrent objects transition to a specified storage class.

awsS3BucketBucketLifecycleConfigurationRulesDetails_noncurrentVersionExpirationInDays :: Lens' AwsS3BucketBucketLifecycleConfigurationRulesDetails (Maybe Int)

The number of days that an object is noncurrent before Amazon S3 can perform the associated action.

awsS3BucketBucketLifecycleConfigurationRulesDetails_expirationDate :: Lens' AwsS3BucketBucketLifecycleConfigurationRulesDetails (Maybe Text)

The date when objects are moved or deleted.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsS3BucketBucketLifecycleConfigurationRulesDetails_id :: Lens' AwsS3BucketBucketLifecycleConfigurationRulesDetails (Maybe Text)

The unique identifier of the rule.

awsS3BucketBucketLifecycleConfigurationRulesDetails_filter :: Lens' AwsS3BucketBucketLifecycleConfigurationRulesDetails (Maybe AwsS3BucketBucketLifecycleConfigurationRulesFilterDetails)

Identifies the objects that a rule applies to.

awsS3BucketBucketLifecycleConfigurationRulesDetails_expirationInDays :: Lens' AwsS3BucketBucketLifecycleConfigurationRulesDetails (Maybe Int)

The length in days of the lifetime for objects that are subject to the rule.

awsS3BucketBucketLifecycleConfigurationRulesDetails_abortIncompleteMultipartUpload :: Lens' AwsS3BucketBucketLifecycleConfigurationRulesDetails (Maybe AwsS3BucketBucketLifecycleConfigurationRulesAbortIncompleteMultipartUploadDetails)

How Amazon S3 responds when a multipart upload is incomplete. Specifically, provides a number of days before Amazon S3 cancels the entire upload.

awsS3BucketBucketLifecycleConfigurationRulesDetails_expiredObjectDeleteMarker :: Lens' AwsS3BucketBucketLifecycleConfigurationRulesDetails (Maybe Bool)

Whether Amazon S3 removes a delete marker that has no noncurrent versions. If set to true, the delete marker is expired. If set to false, the policy takes no action.

If you provide ExpiredObjectDeleteMarker, you cannot provide ExpirationInDays or ExpirationDate.