Amazonka.SecurityHub.Types.AwsS3BucketNotificationConfigurationDetail

Description

Documentation

data AwsS3BucketNotificationConfigurationDetail

Details for an S3 bucket notification configuration.

See: newAwsS3BucketNotificationConfigurationDetail smart constructor.

Constructors

AwsS3BucketNotificationConfigurationDetail'

Fields

destination :: Maybe Text

The ARN of the Lambda function, Amazon SQS queue, or Amazon SNS topic that generates the notification.

events :: Maybe [Text]

The list of events that trigger a notification.

type' :: Maybe Text

Indicates the type of notification. Notifications can be generated using Lambda functions, Amazon SQS queues or Amazon SNS topics.

filter' :: Maybe AwsS3BucketNotificationConfigurationFilter

The filters that determine which S3 buckets generate notifications.

newAwsS3BucketNotificationConfigurationDetail :: AwsS3BucketNotificationConfigurationDetail

Create a value of AwsS3BucketNotificationConfigurationDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destination:AwsS3BucketNotificationConfigurationDetail', awsS3BucketNotificationConfigurationDetail_destination - The ARN of the Lambda function, Amazon SQS queue, or Amazon SNS topic that generates the notification.

$sel:events:AwsS3BucketNotificationConfigurationDetail', awsS3BucketNotificationConfigurationDetail_events - The list of events that trigger a notification.

$sel:type':AwsS3BucketNotificationConfigurationDetail', awsS3BucketNotificationConfigurationDetail_type - Indicates the type of notification. Notifications can be generated using Lambda functions, Amazon SQS queues or Amazon SNS topics.

$sel:filter':AwsS3BucketNotificationConfigurationDetail', awsS3BucketNotificationConfigurationDetail_filter - The filters that determine which S3 buckets generate notifications.

awsS3BucketNotificationConfigurationDetail_destination :: Lens' AwsS3BucketNotificationConfigurationDetail (Maybe Text)

The ARN of the Lambda function, Amazon SQS queue, or Amazon SNS topic that generates the notification.

awsS3BucketNotificationConfigurationDetail_events :: Lens' AwsS3BucketNotificationConfigurationDetail (Maybe [Text])

The list of events that trigger a notification.

awsS3BucketNotificationConfigurationDetail_type :: Lens' AwsS3BucketNotificationConfigurationDetail (Maybe Text)

Indicates the type of notification. Notifications can be generated using Lambda functions, Amazon SQS queues or Amazon SNS topics.

awsS3BucketNotificationConfigurationDetail_filter :: Lens' AwsS3BucketNotificationConfigurationDetail (Maybe AwsS3BucketNotificationConfigurationFilter)

The filters that determine which S3 buckets generate notifications.