Amazonka.SecurityHub.Types.AwsS3BucketWebsiteConfigurationRoutingRuleRedirect

Description

Documentation

data AwsS3BucketWebsiteConfigurationRoutingRuleRedirect

The rules to redirect the request if the condition in Condition is met.

See: newAwsS3BucketWebsiteConfigurationRoutingRuleRedirect smart constructor.

Constructors

AwsS3BucketWebsiteConfigurationRoutingRuleRedirect'

Fields

hostname :: Maybe Text

The host name to use in the redirect request.

protocol :: Maybe Text

The protocol to use to redirect the request. By default, uses the protocol from the original request.

httpRedirectCode :: Maybe Text

The HTTP redirect code to use in the response.

replaceKeyWith :: Maybe Text

The specific object key to use in the redirect request.

Cannot be provided if ReplaceKeyPrefixWith is present.

replaceKeyPrefixWith :: Maybe Text

The object key prefix to use in the redirect request.

Cannot be provided if ReplaceKeyWith is present.

newAwsS3BucketWebsiteConfigurationRoutingRuleRedirect :: AwsS3BucketWebsiteConfigurationRoutingRuleRedirect

Create a value of AwsS3BucketWebsiteConfigurationRoutingRuleRedirect with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostname:AwsS3BucketWebsiteConfigurationRoutingRuleRedirect', awsS3BucketWebsiteConfigurationRoutingRuleRedirect_hostname - The host name to use in the redirect request.

$sel:protocol:AwsS3BucketWebsiteConfigurationRoutingRuleRedirect', awsS3BucketWebsiteConfigurationRoutingRuleRedirect_protocol - The protocol to use to redirect the request. By default, uses the protocol from the original request.

$sel:httpRedirectCode:AwsS3BucketWebsiteConfigurationRoutingRuleRedirect', awsS3BucketWebsiteConfigurationRoutingRuleRedirect_httpRedirectCode - The HTTP redirect code to use in the response.

$sel:replaceKeyWith:AwsS3BucketWebsiteConfigurationRoutingRuleRedirect', awsS3BucketWebsiteConfigurationRoutingRuleRedirect_replaceKeyWith - The specific object key to use in the redirect request.

Cannot be provided if ReplaceKeyPrefixWith is present.

$sel:replaceKeyPrefixWith:AwsS3BucketWebsiteConfigurationRoutingRuleRedirect', awsS3BucketWebsiteConfigurationRoutingRuleRedirect_replaceKeyPrefixWith - The object key prefix to use in the redirect request.

Cannot be provided if ReplaceKeyWith is present.

awsS3BucketWebsiteConfigurationRoutingRuleRedirect_hostname :: Lens' AwsS3BucketWebsiteConfigurationRoutingRuleRedirect (Maybe Text)

The host name to use in the redirect request.

awsS3BucketWebsiteConfigurationRoutingRuleRedirect_protocol :: Lens' AwsS3BucketWebsiteConfigurationRoutingRuleRedirect (Maybe Text)

The protocol to use to redirect the request. By default, uses the protocol from the original request.

awsS3BucketWebsiteConfigurationRoutingRuleRedirect_httpRedirectCode :: Lens' AwsS3BucketWebsiteConfigurationRoutingRuleRedirect (Maybe Text)

The HTTP redirect code to use in the response.

awsS3BucketWebsiteConfigurationRoutingRuleRedirect_replaceKeyWith :: Lens' AwsS3BucketWebsiteConfigurationRoutingRuleRedirect (Maybe Text)

The specific object key to use in the redirect request.

Cannot be provided if ReplaceKeyPrefixWith is present.

awsS3BucketWebsiteConfigurationRoutingRuleRedirect_replaceKeyPrefixWith :: Lens' AwsS3BucketWebsiteConfigurationRoutingRuleRedirect (Maybe Text)

The object key prefix to use in the redirect request.

Cannot be provided if ReplaceKeyWith is present.