Amazonka.SecurityHub.Types.AwsSecurityFinding

Description

Documentation

data AwsSecurityFinding

Provides consistent format for the contents of the Security Hub-aggregated findings. AwsSecurityFinding format enables you to share findings between Amazon Web Services security services and third-party solutions, and security standards checks.

A finding is a potential security issue generated either by Amazon Web Services services or by the integrated third-party solutions and standards checks.

See: newAwsSecurityFinding smart constructor.

Constructors

AwsSecurityFinding'

Fields

productName :: Maybe Text

The name of the product that generated the finding.

Security Hub populates this attribute automatically for each finding. You cannot update it using BatchImportFindings or BatchUpdateFindings. The exception to this is when you use a custom integration.

When you use the Security Hub console to filter findings by product name, you use this attribute.

When you use the Security Hub API to filter findings by product name, you use the aws/securityhub/ProductName attribute under ProductFields.

Security Hub does not synchronize those two attributes.

workflowState :: Maybe WorkflowState

The workflow state of a finding.

criticality :: Maybe Int

The level of importance assigned to the resources associated with the finding.

A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

recordState :: Maybe RecordState

The record state of a finding.

productFields :: Maybe (HashMap Text Text)

A data type where security-findings providers can include additional solution-specific details that aren't part of the defined AwsSecurityFinding format.

Can contain up to 50 key-value pairs. For each key-value pair, the key can contain up to 128 characters, and the value can contain up to 2048 characters.

compliance :: Maybe Compliance

This data type is exclusive to findings that are generated as the result of a check run against a specific rule in a supported security standard, such as CIS Amazon Web Services Foundations. Contains security standard-related finding details.

findingProviderFields :: Maybe FindingProviderFields

In a BatchImportFindings request, finding providers use FindingProviderFields to provide and update their own values for confidence, criticality, related findings, severity, and types.

companyName :: Maybe Text

The name of the company for the product that generated the finding.

Security Hub populates this attribute automatically for each finding. You cannot be updated using BatchImportFindings or BatchUpdateFindings. The exception to this is when you use a custom integration.

When you use the Security Hub console to filter findings by company name, you use this attribute.

When you use the Security Hub API to filter findings by company name, you use the aws/securityhub/CompanyName attribute under ProductFields.

Security Hub does not synchronize those two attributes.

note :: Maybe Note

A user-defined note added to a finding.

process :: Maybe ProcessDetails

The details of process-related information about a finding.

severity :: Maybe Severity

A finding's severity.

types :: Maybe [Text]

One or more finding types in the format of namespace/category/classifier that classify a finding.

Valid namespace values are: Software and Configuration Checks | TTPs | Effects | Unusual Behaviors | Sensitive Data Identifications

action :: Maybe Action

Provides details about an action that affects or that was taken on a resource.

network :: Maybe Network

The details of network-related information about a finding.

relatedFindings :: Maybe [RelatedFinding]

A list of related findings.

firstObservedAt :: Maybe Text

Indicates when the security-findings provider first observed the potential security issue that a finding captured.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

malware :: Maybe [Malware]

A list of malware related to a finding.

confidence :: Maybe Int

A finding's confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify.

Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

remediation :: Maybe Remediation

A data type that describes the remediation options for a finding.

patchSummary :: Maybe PatchSummary

Provides an overview of the patch compliance status for an instance against a selected compliance standard.

vulnerabilities :: Maybe [Vulnerability]

Provides a list of vulnerabilities associated with the findings.

region :: Maybe Text

The Region from which the finding was generated.

Security Hub populates this attribute automatically for each finding. You cannot update it using BatchImportFindings or BatchUpdateFindings.

networkPath :: Maybe [NetworkPathComponent]

Provides information about a network path that is relevant to a finding. Each entry under NetworkPath represents a component of that path.

workflow :: Maybe Workflow

Provides information about the status of the investigation into a finding.

verificationState :: Maybe VerificationState

Indicates the veracity of a finding.

threatIntelIndicators :: Maybe [ThreatIntelIndicator]

Threat intelligence details related to a finding.

sourceUrl :: Maybe Text

A URL that links to a page about the current finding in the security-findings provider's solution.

lastObservedAt :: Maybe Text

Indicates when the security-findings provider most recently observed the potential security issue that a finding captured.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

userDefinedFields :: Maybe (HashMap Text Text)

A list of name/value string pairs associated with the finding. These are custom, user-defined fields added to a finding.

schemaVersion :: Text

The schema version that a finding is formatted for.

id :: Text

The security findings provider-specific identifier for a finding.

productArn :: Text

The ARN generated by Security Hub that uniquely identifies a product that generates findings. This can be the ARN for a third-party product that is integrated with Security Hub, or the ARN for a custom integration.

generatorId :: Text

The identifier for the solution-specific component (a discrete unit of logic) that generated a finding. In various security-findings providers' solutions, this generator can be called a rule, a check, a detector, a plugin, etc.

awsAccountId :: Text

The Amazon Web Services account ID that a finding is generated in.

createdAt :: Text

Indicates when the security-findings provider created the potential security issue that a finding captured.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

updatedAt :: Text

Indicates when the security-findings provider last updated the finding record.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

title :: Text

A finding's title.

In this release, Title is a required property.

description :: Text

A finding's description.

In this release, Description is a required property.

resources :: [Resource]

A set of resource data types that describe the resources that the finding refers to.

newAwsSecurityFinding

Arguments

:: Text
$sel:schemaVersion:AwsSecurityFinding'

-> Text
$sel:id:AwsSecurityFinding'

-> Text
$sel:productArn:AwsSecurityFinding'

-> Text
$sel:generatorId:AwsSecurityFinding'

-> Text
$sel:awsAccountId:AwsSecurityFinding'

-> Text
$sel:createdAt:AwsSecurityFinding'

-> Text
$sel:updatedAt:AwsSecurityFinding'

-> Text
$sel:title:AwsSecurityFinding'

-> Text
$sel:description:AwsSecurityFinding'

-> AwsSecurityFinding

Create a value of AwsSecurityFinding with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productName:AwsSecurityFinding', awsSecurityFinding_productName - The name of the product that generated the finding.

Security Hub populates this attribute automatically for each finding. You cannot update it using BatchImportFindings or BatchUpdateFindings. The exception to this is when you use a custom integration.

When you use the Security Hub console to filter findings by product name, you use this attribute.

When you use the Security Hub API to filter findings by product name, you use the aws/securityhub/ProductName attribute under ProductFields.

Security Hub does not synchronize those two attributes.

$sel:workflowState:AwsSecurityFinding', awsSecurityFinding_workflowState - The workflow state of a finding.

$sel:criticality:AwsSecurityFinding', awsSecurityFinding_criticality - The level of importance assigned to the resources associated with the finding.

A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

$sel:recordState:AwsSecurityFinding', awsSecurityFinding_recordState - The record state of a finding.

$sel:productFields:AwsSecurityFinding', awsSecurityFinding_productFields - A data type where security-findings providers can include additional solution-specific details that aren't part of the defined AwsSecurityFinding format.

Can contain up to 50 key-value pairs. For each key-value pair, the key can contain up to 128 characters, and the value can contain up to 2048 characters.

$sel:compliance:AwsSecurityFinding', awsSecurityFinding_compliance - This data type is exclusive to findings that are generated as the result of a check run against a specific rule in a supported security standard, such as CIS Amazon Web Services Foundations. Contains security standard-related finding details.

$sel:findingProviderFields:AwsSecurityFinding', awsSecurityFinding_findingProviderFields - In a BatchImportFindings request, finding providers use FindingProviderFields to provide and update their own values for confidence, criticality, related findings, severity, and types.

$sel:companyName:AwsSecurityFinding', awsSecurityFinding_companyName - The name of the company for the product that generated the finding.

Security Hub populates this attribute automatically for each finding. You cannot be updated using BatchImportFindings or BatchUpdateFindings. The exception to this is when you use a custom integration.

When you use the Security Hub console to filter findings by company name, you use this attribute.

When you use the Security Hub API to filter findings by company name, you use the aws/securityhub/CompanyName attribute under ProductFields.

Security Hub does not synchronize those two attributes.

$sel:note:AwsSecurityFinding', awsSecurityFinding_note - A user-defined note added to a finding.

$sel:process:AwsSecurityFinding', awsSecurityFinding_process - The details of process-related information about a finding.

$sel:severity:AwsSecurityFinding', awsSecurityFinding_severity - A finding's severity.

$sel:types:AwsSecurityFinding', awsSecurityFinding_types - One or more finding types in the format of namespace/category/classifier that classify a finding.

Valid namespace values are: Software and Configuration Checks | TTPs | Effects | Unusual Behaviors | Sensitive Data Identifications

$sel:action:AwsSecurityFinding', awsSecurityFinding_action - Provides details about an action that affects or that was taken on a resource.

$sel:network:AwsSecurityFinding', awsSecurityFinding_network - The details of network-related information about a finding.

$sel:relatedFindings:AwsSecurityFinding', awsSecurityFinding_relatedFindings - A list of related findings.

$sel:firstObservedAt:AwsSecurityFinding', awsSecurityFinding_firstObservedAt - Indicates when the security-findings provider first observed the potential security issue that a finding captured.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:malware:AwsSecurityFinding', awsSecurityFinding_malware - A list of malware related to a finding.

$sel:confidence:AwsSecurityFinding', awsSecurityFinding_confidence - A finding's confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify.

Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

$sel:remediation:AwsSecurityFinding', awsSecurityFinding_remediation - A data type that describes the remediation options for a finding.

$sel:patchSummary:AwsSecurityFinding', awsSecurityFinding_patchSummary - Provides an overview of the patch compliance status for an instance against a selected compliance standard.

$sel:vulnerabilities:AwsSecurityFinding', awsSecurityFinding_vulnerabilities - Provides a list of vulnerabilities associated with the findings.

$sel:region:AwsSecurityFinding', awsSecurityFinding_region - The Region from which the finding was generated.

Security Hub populates this attribute automatically for each finding. You cannot update it using BatchImportFindings or BatchUpdateFindings.

$sel:networkPath:AwsSecurityFinding', awsSecurityFinding_networkPath - Provides information about a network path that is relevant to a finding. Each entry under NetworkPath represents a component of that path.

$sel:workflow:AwsSecurityFinding', awsSecurityFinding_workflow - Provides information about the status of the investigation into a finding.

$sel:verificationState:AwsSecurityFinding', awsSecurityFinding_verificationState - Indicates the veracity of a finding.

$sel:threatIntelIndicators:AwsSecurityFinding', awsSecurityFinding_threatIntelIndicators - Threat intelligence details related to a finding.

$sel:sourceUrl:AwsSecurityFinding', awsSecurityFinding_sourceUrl - A URL that links to a page about the current finding in the security-findings provider's solution.

$sel:lastObservedAt:AwsSecurityFinding', awsSecurityFinding_lastObservedAt - Indicates when the security-findings provider most recently observed the potential security issue that a finding captured.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:userDefinedFields:AwsSecurityFinding', awsSecurityFinding_userDefinedFields - A list of name/value string pairs associated with the finding. These are custom, user-defined fields added to a finding.

$sel:schemaVersion:AwsSecurityFinding', awsSecurityFinding_schemaVersion - The schema version that a finding is formatted for.

$sel:id:AwsSecurityFinding', awsSecurityFinding_id - The security findings provider-specific identifier for a finding.

$sel:productArn:AwsSecurityFinding', awsSecurityFinding_productArn - The ARN generated by Security Hub that uniquely identifies a product that generates findings. This can be the ARN for a third-party product that is integrated with Security Hub, or the ARN for a custom integration.

$sel:generatorId:AwsSecurityFinding', awsSecurityFinding_generatorId - The identifier for the solution-specific component (a discrete unit of logic) that generated a finding. In various security-findings providers' solutions, this generator can be called a rule, a check, a detector, a plugin, etc.

$sel:awsAccountId:AwsSecurityFinding', awsSecurityFinding_awsAccountId - The Amazon Web Services account ID that a finding is generated in.

$sel:createdAt:AwsSecurityFinding', awsSecurityFinding_createdAt - Indicates when the security-findings provider created the potential security issue that a finding captured.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:updatedAt:AwsSecurityFinding', awsSecurityFinding_updatedAt - Indicates when the security-findings provider last updated the finding record.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:title:AwsSecurityFinding', awsSecurityFinding_title - A finding's title.

In this release, Title is a required property.

$sel:description:AwsSecurityFinding', awsSecurityFinding_description - A finding's description.

In this release, Description is a required property.

$sel:resources:AwsSecurityFinding', awsSecurityFinding_resources - A set of resource data types that describe the resources that the finding refers to.

awsSecurityFinding_productName :: Lens' AwsSecurityFinding (Maybe Text)

The name of the product that generated the finding.

Security Hub populates this attribute automatically for each finding. You cannot update it using BatchImportFindings or BatchUpdateFindings. The exception to this is when you use a custom integration.

When you use the Security Hub console to filter findings by product name, you use this attribute.

When you use the Security Hub API to filter findings by product name, you use the aws/securityhub/ProductName attribute under ProductFields.

Security Hub does not synchronize those two attributes.

awsSecurityFinding_workflowState :: Lens' AwsSecurityFinding (Maybe WorkflowState)

The workflow state of a finding.

awsSecurityFinding_criticality :: Lens' AwsSecurityFinding (Maybe Int)

The level of importance assigned to the resources associated with the finding.

A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

awsSecurityFinding_recordState :: Lens' AwsSecurityFinding (Maybe RecordState)

The record state of a finding.

awsSecurityFinding_productFields :: Lens' AwsSecurityFinding (Maybe (HashMap Text Text))

A data type where security-findings providers can include additional solution-specific details that aren't part of the defined AwsSecurityFinding format.

Can contain up to 50 key-value pairs. For each key-value pair, the key can contain up to 128 characters, and the value can contain up to 2048 characters.

awsSecurityFinding_compliance :: Lens' AwsSecurityFinding (Maybe Compliance)

This data type is exclusive to findings that are generated as the result of a check run against a specific rule in a supported security standard, such as CIS Amazon Web Services Foundations. Contains security standard-related finding details.

awsSecurityFinding_findingProviderFields :: Lens' AwsSecurityFinding (Maybe FindingProviderFields)

In a BatchImportFindings request, finding providers use FindingProviderFields to provide and update their own values for confidence, criticality, related findings, severity, and types.

awsSecurityFinding_companyName :: Lens' AwsSecurityFinding (Maybe Text)

The name of the company for the product that generated the finding.

Security Hub populates this attribute automatically for each finding. You cannot be updated using BatchImportFindings or BatchUpdateFindings. The exception to this is when you use a custom integration.

When you use the Security Hub console to filter findings by company name, you use this attribute.

When you use the Security Hub API to filter findings by company name, you use the aws/securityhub/CompanyName attribute under ProductFields.

Security Hub does not synchronize those two attributes.

awsSecurityFinding_note :: Lens' AwsSecurityFinding (Maybe Note)

A user-defined note added to a finding.

awsSecurityFinding_process :: Lens' AwsSecurityFinding (Maybe ProcessDetails)

The details of process-related information about a finding.

awsSecurityFinding_severity :: Lens' AwsSecurityFinding (Maybe Severity)

A finding's severity.

awsSecurityFinding_types :: Lens' AwsSecurityFinding (Maybe [Text])

One or more finding types in the format of namespace/category/classifier that classify a finding.

Valid namespace values are: Software and Configuration Checks | TTPs | Effects | Unusual Behaviors | Sensitive Data Identifications

awsSecurityFinding_action :: Lens' AwsSecurityFinding (Maybe Action)

Provides details about an action that affects or that was taken on a resource.

awsSecurityFinding_network :: Lens' AwsSecurityFinding (Maybe Network)

The details of network-related information about a finding.

awsSecurityFinding_relatedFindings :: Lens' AwsSecurityFinding (Maybe [RelatedFinding])

A list of related findings.

awsSecurityFinding_firstObservedAt :: Lens' AwsSecurityFinding (Maybe Text)

Indicates when the security-findings provider first observed the potential security issue that a finding captured.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsSecurityFinding_malware :: Lens' AwsSecurityFinding (Maybe [Malware])

A list of malware related to a finding.

awsSecurityFinding_confidence :: Lens' AwsSecurityFinding (Maybe Int)

A finding's confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify.

Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

awsSecurityFinding_remediation :: Lens' AwsSecurityFinding (Maybe Remediation)

A data type that describes the remediation options for a finding.

awsSecurityFinding_patchSummary :: Lens' AwsSecurityFinding (Maybe PatchSummary)

Provides an overview of the patch compliance status for an instance against a selected compliance standard.

awsSecurityFinding_vulnerabilities :: Lens' AwsSecurityFinding (Maybe [Vulnerability])

Provides a list of vulnerabilities associated with the findings.

awsSecurityFinding_region :: Lens' AwsSecurityFinding (Maybe Text)

The Region from which the finding was generated.

Security Hub populates this attribute automatically for each finding. You cannot update it using BatchImportFindings or BatchUpdateFindings.

awsSecurityFinding_networkPath :: Lens' AwsSecurityFinding (Maybe [NetworkPathComponent])

Provides information about a network path that is relevant to a finding. Each entry under NetworkPath represents a component of that path.

awsSecurityFinding_workflow :: Lens' AwsSecurityFinding (Maybe Workflow)

Provides information about the status of the investigation into a finding.

awsSecurityFinding_verificationState :: Lens' AwsSecurityFinding (Maybe VerificationState)

Indicates the veracity of a finding.

awsSecurityFinding_threatIntelIndicators :: Lens' AwsSecurityFinding (Maybe [ThreatIntelIndicator])

Threat intelligence details related to a finding.

awsSecurityFinding_sourceUrl :: Lens' AwsSecurityFinding (Maybe Text)

A URL that links to a page about the current finding in the security-findings provider's solution.

awsSecurityFinding_lastObservedAt :: Lens' AwsSecurityFinding (Maybe Text)

Indicates when the security-findings provider most recently observed the potential security issue that a finding captured.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsSecurityFinding_userDefinedFields :: Lens' AwsSecurityFinding (Maybe (HashMap Text Text))

A list of name/value string pairs associated with the finding. These are custom, user-defined fields added to a finding.

awsSecurityFinding_schemaVersion :: Lens' AwsSecurityFinding Text

The schema version that a finding is formatted for.

awsSecurityFinding_id :: Lens' AwsSecurityFinding Text

The security findings provider-specific identifier for a finding.

awsSecurityFinding_productArn :: Lens' AwsSecurityFinding Text

The ARN generated by Security Hub that uniquely identifies a product that generates findings. This can be the ARN for a third-party product that is integrated with Security Hub, or the ARN for a custom integration.

awsSecurityFinding_generatorId :: Lens' AwsSecurityFinding Text

The identifier for the solution-specific component (a discrete unit of logic) that generated a finding. In various security-findings providers' solutions, this generator can be called a rule, a check, a detector, a plugin, etc.

awsSecurityFinding_awsAccountId :: Lens' AwsSecurityFinding Text

The Amazon Web Services account ID that a finding is generated in.

awsSecurityFinding_createdAt :: Lens' AwsSecurityFinding Text

Indicates when the security-findings provider created the potential security issue that a finding captured.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsSecurityFinding_updatedAt :: Lens' AwsSecurityFinding Text

Indicates when the security-findings provider last updated the finding record.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

awsSecurityFinding_title :: Lens' AwsSecurityFinding Text

A finding's title.

In this release, Title is a required property.

awsSecurityFinding_description :: Lens' AwsSecurityFinding Text

A finding's description.

In this release, Description is a required property.

awsSecurityFinding_resources :: Lens' AwsSecurityFinding [Resource]

A set of resource data types that describe the resources that the finding refers to.