Amazonka.SecurityHub.Types.AwsSsmComplianceSummary

Description

Documentation

data AwsSsmComplianceSummary

Provides the details about the compliance status for a patch.

See: newAwsSsmComplianceSummary smart constructor.

Constructors

AwsSsmComplianceSummary'

Fields

status :: Maybe Text

The current patch compliance status.

The possible status values are:

COMPLIANT

NON_COMPLIANT

UNSPECIFIED_DATA

nonCompliantLowCount :: Maybe Int

For the patches that are noncompliant, the number that have a severity of LOW.

compliantHighCount :: Maybe Int

For the patches that are compliant, the number that have a severity of HIGH.

nonCompliantUnspecifiedCount :: Maybe Int

For the patches that are noncompliant, the number that have a severity of UNSPECIFIED.

executionType :: Maybe Text

The type of execution that was used determine compliance.

compliantInformationalCount :: Maybe Int

For the patches that are compliant, the number that have a severity of INFORMATIONAL.

nonCompliantHighCount :: Maybe Int

For the patches that are noncompliant, the number that have a severity of HIGH.

nonCompliantMediumCount :: Maybe Int

For the patches that are noncompliant, the number that have a severity of MEDIUM.

overallSeverity :: Maybe Text

The highest severity for the patches.

compliantCriticalCount :: Maybe Int

For the patches that are compliant, the number that have a severity of CRITICAL.

nonCompliantInformationalCount :: Maybe Int

For the patches that are noncompliant, the number that have a severity of INFORMATIONAL.

complianceType :: Maybe Text

The type of resource for which the compliance was determined. For AwsSsmPatchCompliance, ComplianceType is Patch.

compliantUnspecifiedCount :: Maybe Int

For the patches that are compliant, the number that have a severity of UNSPECIFIED.

nonCompliantCriticalCount :: Maybe Int

For the patch items that are noncompliant, the number of items that have a severity of CRITICAL.

patchBaselineId :: Maybe Text

The identifier of the patch baseline. The patch baseline lists the patches that are approved for installation.

compliantLowCount :: Maybe Int

For the patches that are compliant, the number that have a severity of LOW.

patchGroup :: Maybe Text

The identifier of the patch group for which compliance was determined. A patch group uses tags to group EC2 instances that should have the same patch compliance.

compliantMediumCount :: Maybe Int

For the patches that are compliant, the number that have a severity of MEDIUM.

newAwsSsmComplianceSummary :: AwsSsmComplianceSummary

Create a value of AwsSsmComplianceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AwsSsmComplianceSummary', awsSsmComplianceSummary_status - The current patch compliance status.

The possible status values are:

COMPLIANT

NON_COMPLIANT

UNSPECIFIED_DATA

$sel:nonCompliantLowCount:AwsSsmComplianceSummary', awsSsmComplianceSummary_nonCompliantLowCount - For the patches that are noncompliant, the number that have a severity of LOW.

$sel:compliantHighCount:AwsSsmComplianceSummary', awsSsmComplianceSummary_compliantHighCount - For the patches that are compliant, the number that have a severity of HIGH.

$sel:nonCompliantUnspecifiedCount:AwsSsmComplianceSummary', awsSsmComplianceSummary_nonCompliantUnspecifiedCount - For the patches that are noncompliant, the number that have a severity of UNSPECIFIED.

$sel:executionType:AwsSsmComplianceSummary', awsSsmComplianceSummary_executionType - The type of execution that was used determine compliance.

$sel:compliantInformationalCount:AwsSsmComplianceSummary', awsSsmComplianceSummary_compliantInformationalCount - For the patches that are compliant, the number that have a severity of INFORMATIONAL.

$sel:nonCompliantHighCount:AwsSsmComplianceSummary', awsSsmComplianceSummary_nonCompliantHighCount - For the patches that are noncompliant, the number that have a severity of HIGH.

$sel:nonCompliantMediumCount:AwsSsmComplianceSummary', awsSsmComplianceSummary_nonCompliantMediumCount - For the patches that are noncompliant, the number that have a severity of MEDIUM.

$sel:overallSeverity:AwsSsmComplianceSummary', awsSsmComplianceSummary_overallSeverity - The highest severity for the patches.

$sel:compliantCriticalCount:AwsSsmComplianceSummary', awsSsmComplianceSummary_compliantCriticalCount - For the patches that are compliant, the number that have a severity of CRITICAL.

$sel:nonCompliantInformationalCount:AwsSsmComplianceSummary', awsSsmComplianceSummary_nonCompliantInformationalCount - For the patches that are noncompliant, the number that have a severity of INFORMATIONAL.

$sel:complianceType:AwsSsmComplianceSummary', awsSsmComplianceSummary_complianceType - The type of resource for which the compliance was determined. For AwsSsmPatchCompliance, ComplianceType is Patch.

$sel:compliantUnspecifiedCount:AwsSsmComplianceSummary', awsSsmComplianceSummary_compliantUnspecifiedCount - For the patches that are compliant, the number that have a severity of UNSPECIFIED.

$sel:nonCompliantCriticalCount:AwsSsmComplianceSummary', awsSsmComplianceSummary_nonCompliantCriticalCount - For the patch items that are noncompliant, the number of items that have a severity of CRITICAL.

$sel:patchBaselineId:AwsSsmComplianceSummary', awsSsmComplianceSummary_patchBaselineId - The identifier of the patch baseline. The patch baseline lists the patches that are approved for installation.

$sel:compliantLowCount:AwsSsmComplianceSummary', awsSsmComplianceSummary_compliantLowCount - For the patches that are compliant, the number that have a severity of LOW.

$sel:patchGroup:AwsSsmComplianceSummary', awsSsmComplianceSummary_patchGroup - The identifier of the patch group for which compliance was determined. A patch group uses tags to group EC2 instances that should have the same patch compliance.

$sel:compliantMediumCount:AwsSsmComplianceSummary', awsSsmComplianceSummary_compliantMediumCount - For the patches that are compliant, the number that have a severity of MEDIUM.

awsSsmComplianceSummary_status :: Lens' AwsSsmComplianceSummary (Maybe Text)

The current patch compliance status.

The possible status values are:

COMPLIANT

NON_COMPLIANT

UNSPECIFIED_DATA

awsSsmComplianceSummary_nonCompliantLowCount :: Lens' AwsSsmComplianceSummary (Maybe Int)

For the patches that are noncompliant, the number that have a severity of LOW.

awsSsmComplianceSummary_compliantHighCount :: Lens' AwsSsmComplianceSummary (Maybe Int)

For the patches that are compliant, the number that have a severity of HIGH.

awsSsmComplianceSummary_nonCompliantUnspecifiedCount :: Lens' AwsSsmComplianceSummary (Maybe Int)

For the patches that are noncompliant, the number that have a severity of UNSPECIFIED.

awsSsmComplianceSummary_executionType :: Lens' AwsSsmComplianceSummary (Maybe Text)

The type of execution that was used determine compliance.

awsSsmComplianceSummary_compliantInformationalCount :: Lens' AwsSsmComplianceSummary (Maybe Int)

For the patches that are compliant, the number that have a severity of INFORMATIONAL.

awsSsmComplianceSummary_nonCompliantHighCount :: Lens' AwsSsmComplianceSummary (Maybe Int)

For the patches that are noncompliant, the number that have a severity of HIGH.

awsSsmComplianceSummary_nonCompliantMediumCount :: Lens' AwsSsmComplianceSummary (Maybe Int)

For the patches that are noncompliant, the number that have a severity of MEDIUM.

awsSsmComplianceSummary_overallSeverity :: Lens' AwsSsmComplianceSummary (Maybe Text)

The highest severity for the patches.

awsSsmComplianceSummary_compliantCriticalCount :: Lens' AwsSsmComplianceSummary (Maybe Int)

For the patches that are compliant, the number that have a severity of CRITICAL.

awsSsmComplianceSummary_nonCompliantInformationalCount :: Lens' AwsSsmComplianceSummary (Maybe Int)

For the patches that are noncompliant, the number that have a severity of INFORMATIONAL.

awsSsmComplianceSummary_complianceType :: Lens' AwsSsmComplianceSummary (Maybe Text)

The type of resource for which the compliance was determined. For AwsSsmPatchCompliance, ComplianceType is Patch.

awsSsmComplianceSummary_compliantUnspecifiedCount :: Lens' AwsSsmComplianceSummary (Maybe Int)

For the patches that are compliant, the number that have a severity of UNSPECIFIED.

awsSsmComplianceSummary_nonCompliantCriticalCount :: Lens' AwsSsmComplianceSummary (Maybe Int)

For the patch items that are noncompliant, the number of items that have a severity of CRITICAL.

awsSsmComplianceSummary_patchBaselineId :: Lens' AwsSsmComplianceSummary (Maybe Text)

The identifier of the patch baseline. The patch baseline lists the patches that are approved for installation.

awsSsmComplianceSummary_compliantLowCount :: Lens' AwsSsmComplianceSummary (Maybe Int)

For the patches that are compliant, the number that have a severity of LOW.

awsSsmComplianceSummary_patchGroup :: Lens' AwsSsmComplianceSummary (Maybe Text)

The identifier of the patch group for which compliance was determined. A patch group uses tags to group EC2 instances that should have the same patch compliance.

awsSsmComplianceSummary_compliantMediumCount :: Lens' AwsSsmComplianceSummary (Maybe Int)

For the patches that are compliant, the number that have a severity of MEDIUM.