Amazonka.SecurityHub.Types.AwsWafWebAclRule

Description

Documentation

data AwsWafWebAclRule

Details for a rule in an WAF WebACL.

See: newAwsWafWebAclRule smart constructor.

Constructors

AwsWafWebAclRule'

Fields

priority :: Maybe Int

Specifies the order in which the rules in a WebACL are evaluated. Rules with a lower value for Priority are evaluated before rules with a higher value. The value must be a unique integer. If you add multiple rules to a WebACL, the values do not need to be consecutive.

overrideAction :: Maybe WafOverrideAction

Use the OverrideAction to test your RuleGroup.

Any rule in a RuleGroup can potentially block a request. If you set the OverrideAction to None, the RuleGroup blocks a request if any individual rule in the RuleGroup matches the request and is configured to block that request.

However, if you first want to test the RuleGroup, set the OverrideAction to Count. The RuleGroup then overrides any block action specified by individual rules contained within the group. Instead of blocking matching requests, those requests are counted.

ActivatedRule|OverrideAction applies only when updating or adding a RuleGroup to a WebACL. In this case you do not use ActivatedRule|Action. For all other update requests, ActivatedRule|Action is used instead of ActivatedRule|OverrideAction.

ruleId :: Maybe Text

The identifier for a rule.

action :: Maybe WafAction

Specifies the action that CloudFront or WAF takes when a web request matches the conditions in the rule.

excludedRules :: Maybe [WafExcludedRule]

Rules to exclude from a rule group.

type' :: Maybe Text

The rule type.

Valid values: REGULAR | RATE_BASED | GROUP

The default is REGULAR.

newAwsWafWebAclRule :: AwsWafWebAclRule

Create a value of AwsWafWebAclRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:priority:AwsWafWebAclRule', awsWafWebAclRule_priority - Specifies the order in which the rules in a WebACL are evaluated. Rules with a lower value for Priority are evaluated before rules with a higher value. The value must be a unique integer. If you add multiple rules to a WebACL, the values do not need to be consecutive.

$sel:overrideAction:AwsWafWebAclRule', awsWafWebAclRule_overrideAction - Use the OverrideAction to test your RuleGroup.

Any rule in a RuleGroup can potentially block a request. If you set the OverrideAction to None, the RuleGroup blocks a request if any individual rule in the RuleGroup matches the request and is configured to block that request.

However, if you first want to test the RuleGroup, set the OverrideAction to Count. The RuleGroup then overrides any block action specified by individual rules contained within the group. Instead of blocking matching requests, those requests are counted.

ActivatedRule|OverrideAction applies only when updating or adding a RuleGroup to a WebACL. In this case you do not use ActivatedRule|Action. For all other update requests, ActivatedRule|Action is used instead of ActivatedRule|OverrideAction.

$sel:ruleId:AwsWafWebAclRule', awsWafWebAclRule_ruleId - The identifier for a rule.

$sel:action:AwsWafWebAclRule', awsWafWebAclRule_action - Specifies the action that CloudFront or WAF takes when a web request matches the conditions in the rule.

$sel:excludedRules:AwsWafWebAclRule', awsWafWebAclRule_excludedRules - Rules to exclude from a rule group.

$sel:type':AwsWafWebAclRule', awsWafWebAclRule_type - The rule type.

Valid values: REGULAR | RATE_BASED | GROUP

The default is REGULAR.

awsWafWebAclRule_priority :: Lens' AwsWafWebAclRule (Maybe Int)

Specifies the order in which the rules in a WebACL are evaluated. Rules with a lower value for Priority are evaluated before rules with a higher value. The value must be a unique integer. If you add multiple rules to a WebACL, the values do not need to be consecutive.

awsWafWebAclRule_overrideAction :: Lens' AwsWafWebAclRule (Maybe WafOverrideAction)

Use the OverrideAction to test your RuleGroup.

Any rule in a RuleGroup can potentially block a request. If you set the OverrideAction to None, the RuleGroup blocks a request if any individual rule in the RuleGroup matches the request and is configured to block that request.

However, if you first want to test the RuleGroup, set the OverrideAction to Count. The RuleGroup then overrides any block action specified by individual rules contained within the group. Instead of blocking matching requests, those requests are counted.

ActivatedRule|OverrideAction applies only when updating or adding a RuleGroup to a WebACL. In this case you do not use ActivatedRule|Action. For all other update requests, ActivatedRule|Action is used instead of ActivatedRule|OverrideAction.

awsWafWebAclRule_ruleId :: Lens' AwsWafWebAclRule (Maybe Text)

The identifier for a rule.

awsWafWebAclRule_action :: Lens' AwsWafWebAclRule (Maybe WafAction)

Specifies the action that CloudFront or WAF takes when a web request matches the conditions in the rule.

awsWafWebAclRule_excludedRules :: Lens' AwsWafWebAclRule (Maybe [WafExcludedRule])

Rules to exclude from a rule group.

awsWafWebAclRule_type :: Lens' AwsWafWebAclRule (Maybe Text)

The rule type.

Valid values: REGULAR | RATE_BASED | GROUP

The default is REGULAR.