Amazonka.SecurityHub.Types.Cvss

Description

Documentation

data Cvss

CVSS scores from the advisory related to the vulnerability.

See: newCvss smart constructor.

Constructors

Cvss'

Fields

adjustments :: Maybe [Adjustment]

Adjustments to the CVSS metrics.

baseVector :: Maybe Text

The base scoring vector for the CVSS score.

version :: Maybe Text

The version of CVSS for the CVSS score.

source :: Maybe Text

The origin of the original CVSS score and vector.

baseScore :: Maybe Double

The base CVSS score.

newCvss :: Cvss

Create a value of Cvss with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:adjustments:Cvss', cvss_adjustments - Adjustments to the CVSS metrics.

$sel:baseVector:Cvss', cvss_baseVector - The base scoring vector for the CVSS score.

$sel:version:Cvss', cvss_version - The version of CVSS for the CVSS score.

$sel:source:Cvss', cvss_source - The origin of the original CVSS score and vector.

$sel:baseScore:Cvss', cvss_baseScore - The base CVSS score.

cvss_adjustments :: Lens' Cvss (Maybe [Adjustment])

Adjustments to the CVSS metrics.

cvss_baseVector :: Lens' Cvss (Maybe Text)

The base scoring vector for the CVSS score.

cvss_version :: Lens' Cvss (Maybe Text)

The version of CVSS for the CVSS score.

cvss_source :: Lens' Cvss (Maybe Text)

The origin of the original CVSS score and vector.

cvss_baseScore :: Lens' Cvss (Maybe Double)

The base CVSS score.