Amazonka.SecurityHub.Types.Member

Description

Documentation

data Member

The details about a member account.

See: newMember smart constructor.

Constructors

Member'

Fields

email :: Maybe Text

The email address of the member account.

invitedAt :: Maybe POSIX

A timestamp for the date and time when the invitation was sent to the member account.

administratorId :: Maybe Text

The Amazon Web Services account ID of the Security Hub administrator account associated with this member account.

masterId :: Maybe Text

This is replaced by AdministratorID.

The Amazon Web Services account ID of the Security Hub administrator account associated with this member account.

accountId :: Maybe Text

The Amazon Web Services account ID of the member account.

memberStatus :: Maybe Text

The status of the relationship between the member account and its administrator account.

The status can have one of the following values:

CREATED - Indicates that the administrator account added the member account, but has not yet invited the member account.

INVITED - Indicates that the administrator account invited the member account. The member account has not yet responded to the invitation.

ENABLED - Indicates that the member account is currently active. For manually invited member accounts, indicates that the member account accepted the invitation.

REMOVED - Indicates that the administrator account disassociated the member account.

RESIGNED - Indicates that the member account disassociated themselves from the administrator account.

DELETED - Indicates that the administrator account deleted the member account.

ACCOUNT_SUSPENDED - Indicates that an organization account was suspended from Amazon Web Services at the same time that the administrator account tried to enable the organization account as a member account.

updatedAt :: Maybe POSIX

The timestamp for the date and time when the member account was updated.

newMember :: Member

Create a value of Member with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:Member', member_email - The email address of the member account.

$sel:invitedAt:Member', member_invitedAt - A timestamp for the date and time when the invitation was sent to the member account.

$sel:administratorId:Member', member_administratorId - The Amazon Web Services account ID of the Security Hub administrator account associated with this member account.

$sel:masterId:Member', member_masterId - This is replaced by AdministratorID.

The Amazon Web Services account ID of the Security Hub administrator account associated with this member account.

$sel:accountId:Member', member_accountId - The Amazon Web Services account ID of the member account.

$sel:memberStatus:Member', member_memberStatus - The status of the relationship between the member account and its administrator account.

The status can have one of the following values:

CREATED - Indicates that the administrator account added the member account, but has not yet invited the member account.

INVITED - Indicates that the administrator account invited the member account. The member account has not yet responded to the invitation.

ENABLED - Indicates that the member account is currently active. For manually invited member accounts, indicates that the member account accepted the invitation.

REMOVED - Indicates that the administrator account disassociated the member account.

RESIGNED - Indicates that the member account disassociated themselves from the administrator account.

DELETED - Indicates that the administrator account deleted the member account.

ACCOUNT_SUSPENDED - Indicates that an organization account was suspended from Amazon Web Services at the same time that the administrator account tried to enable the organization account as a member account.

$sel:updatedAt:Member', member_updatedAt - The timestamp for the date and time when the member account was updated.

member_email :: Lens' Member (Maybe Text)

The email address of the member account.

member_invitedAt :: Lens' Member (Maybe UTCTime)

A timestamp for the date and time when the invitation was sent to the member account.

member_administratorId :: Lens' Member (Maybe Text)

The Amazon Web Services account ID of the Security Hub administrator account associated with this member account.

member_masterId :: Lens' Member (Maybe Text)

This is replaced by AdministratorID.

The Amazon Web Services account ID of the Security Hub administrator account associated with this member account.

member_accountId :: Lens' Member (Maybe Text)

The Amazon Web Services account ID of the member account.

member_memberStatus :: Lens' Member (Maybe Text)

The status of the relationship between the member account and its administrator account.

The status can have one of the following values:

CREATED - Indicates that the administrator account added the member account, but has not yet invited the member account.

INVITED - Indicates that the administrator account invited the member account. The member account has not yet responded to the invitation.

ENABLED - Indicates that the member account is currently active. For manually invited member accounts, indicates that the member account accepted the invitation.

REMOVED - Indicates that the administrator account disassociated the member account.

RESIGNED - Indicates that the member account disassociated themselves from the administrator account.

DELETED - Indicates that the administrator account deleted the member account.

ACCOUNT_SUSPENDED - Indicates that an organization account was suspended from Amazon Web Services at the same time that the administrator account tried to enable the organization account as a member account.

member_updatedAt :: Lens' Member (Maybe UTCTime)

The timestamp for the date and time when the member account was updated.