Amazonka.SecurityHub.Types.Network

Description

Documentation

data Network

The details of network-related information about a finding.

See: newNetwork smart constructor.

Constructors

Network'

Fields

destinationDomain :: Maybe Text

The destination domain of network-related information about a finding.

sourcePort :: Maybe Int

The source port of network-related information about a finding.

openPortRange :: Maybe PortRange

The range of open ports that is present on the network.

sourceMac :: Maybe Text

The source media access control (MAC) address of network-related information about a finding.

direction :: Maybe NetworkDirection

The direction of network traffic associated with a finding.

destinationIpV6 :: Maybe Text

The destination IPv6 address of network-related information about a finding.

protocol :: Maybe Text

The protocol of network-related information about a finding.

sourceIpV6 :: Maybe Text

The source IPv6 address of network-related information about a finding.

destinationIpV4 :: Maybe Text

The destination IPv4 address of network-related information about a finding.

sourceDomain :: Maybe Text

The source domain of network-related information about a finding.

destinationPort :: Maybe Int

The destination port of network-related information about a finding.

sourceIpV4 :: Maybe Text

The source IPv4 address of network-related information about a finding.

newNetwork :: Network

Create a value of Network with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationDomain:Network', network_destinationDomain - The destination domain of network-related information about a finding.

$sel:sourcePort:Network', network_sourcePort - The source port of network-related information about a finding.

$sel:openPortRange:Network', network_openPortRange - The range of open ports that is present on the network.

$sel:sourceMac:Network', network_sourceMac - The source media access control (MAC) address of network-related information about a finding.

$sel:direction:Network', network_direction - The direction of network traffic associated with a finding.

$sel:destinationIpV6:Network', network_destinationIpV6 - The destination IPv6 address of network-related information about a finding.

$sel:protocol:Network', network_protocol - The protocol of network-related information about a finding.

$sel:sourceIpV6:Network', network_sourceIpV6 - The source IPv6 address of network-related information about a finding.

$sel:destinationIpV4:Network', network_destinationIpV4 - The destination IPv4 address of network-related information about a finding.

$sel:sourceDomain:Network', network_sourceDomain - The source domain of network-related information about a finding.

$sel:destinationPort:Network', network_destinationPort - The destination port of network-related information about a finding.

$sel:sourceIpV4:Network', network_sourceIpV4 - The source IPv4 address of network-related information about a finding.

network_destinationDomain :: Lens' Network (Maybe Text)

The destination domain of network-related information about a finding.

network_sourcePort :: Lens' Network (Maybe Int)

The source port of network-related information about a finding.

network_openPortRange :: Lens' Network (Maybe PortRange)

The range of open ports that is present on the network.

network_sourceMac :: Lens' Network (Maybe Text)

The source media access control (MAC) address of network-related information about a finding.

network_direction :: Lens' Network (Maybe NetworkDirection)

The direction of network traffic associated with a finding.

network_destinationIpV6 :: Lens' Network (Maybe Text)

The destination IPv6 address of network-related information about a finding.

network_protocol :: Lens' Network (Maybe Text)

The protocol of network-related information about a finding.

network_sourceIpV6 :: Lens' Network (Maybe Text)

The source IPv6 address of network-related information about a finding.

network_destinationIpV4 :: Lens' Network (Maybe Text)

The destination IPv4 address of network-related information about a finding.

network_sourceDomain :: Lens' Network (Maybe Text)

The source domain of network-related information about a finding.

network_destinationPort :: Lens' Network (Maybe Int)

The destination port of network-related information about a finding.

network_sourceIpV4 :: Lens' Network (Maybe Text)

The source IPv4 address of network-related information about a finding.