Amazonka.SecurityHub.Types.NetworkPathComponent

Description

Documentation

data NetworkPathComponent

Information about a network path component.

See: newNetworkPathComponent smart constructor.

Constructors

NetworkPathComponent'

Fields

componentType :: Maybe Text

The type of component.

ingress :: Maybe NetworkHeader

Information about the component that comes before the current node in the network path.

componentId :: Maybe Text

The identifier of a component in the network path.

egress :: Maybe NetworkHeader

Information about the component that comes after the current component in the network path.

newNetworkPathComponent :: NetworkPathComponent

Create a value of NetworkPathComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:componentType:NetworkPathComponent', networkPathComponent_componentType - The type of component.

$sel:ingress:NetworkPathComponent', networkPathComponent_ingress - Information about the component that comes before the current node in the network path.

$sel:componentId:NetworkPathComponent', networkPathComponent_componentId - The identifier of a component in the network path.

$sel:egress:NetworkPathComponent', networkPathComponent_egress - Information about the component that comes after the current component in the network path.

networkPathComponent_componentType :: Lens' NetworkPathComponent (Maybe Text)

The type of component.

networkPathComponent_ingress :: Lens' NetworkPathComponent (Maybe NetworkHeader)

Information about the component that comes before the current node in the network path.

networkPathComponent_componentId :: Lens' NetworkPathComponent (Maybe Text)

The identifier of a component in the network path.

networkPathComponent_egress :: Lens' NetworkPathComponent (Maybe NetworkHeader)

Information about the component that comes after the current component in the network path.