Amazonka.SecurityHub.Types.Occurrences

Description

Documentation

data Occurrences

The detected occurrences of sensitive data.

See: newOccurrences smart constructor.

Constructors

Occurrences'

Fields

lineRanges :: Maybe [Range]

Occurrences of sensitive data detected in a non-binary text file or a Microsoft Word file. Non-binary text files include files such as HTML, XML, JSON, and TXT files.

cells :: Maybe [Cell]

Occurrences of sensitive data detected in Microsoft Excel workbooks, comma-separated value (CSV) files, or tab-separated value (TSV) files.

pages :: Maybe [Page]

Occurrences of sensitive data in an Adobe Portable Document Format (PDF) file.

records :: Maybe [Record]

Occurrences of sensitive data in an Apache Avro object container or an Apache Parquet file.

offsetRanges :: Maybe [Range]

Occurrences of sensitive data detected in a binary text file.

newOccurrences :: Occurrences

Create a value of Occurrences with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lineRanges:Occurrences', occurrences_lineRanges - Occurrences of sensitive data detected in a non-binary text file or a Microsoft Word file. Non-binary text files include files such as HTML, XML, JSON, and TXT files.

$sel:cells:Occurrences', occurrences_cells - Occurrences of sensitive data detected in Microsoft Excel workbooks, comma-separated value (CSV) files, or tab-separated value (TSV) files.

$sel:pages:Occurrences', occurrences_pages - Occurrences of sensitive data in an Adobe Portable Document Format (PDF) file.

$sel:records:Occurrences', occurrences_records - Occurrences of sensitive data in an Apache Avro object container or an Apache Parquet file.

$sel:offsetRanges:Occurrences', occurrences_offsetRanges - Occurrences of sensitive data detected in a binary text file.

occurrences_lineRanges :: Lens' Occurrences (Maybe [Range])

Occurrences of sensitive data detected in a non-binary text file or a Microsoft Word file. Non-binary text files include files such as HTML, XML, JSON, and TXT files.

occurrences_cells :: Lens' Occurrences (Maybe [Cell])

Occurrences of sensitive data detected in Microsoft Excel workbooks, comma-separated value (CSV) files, or tab-separated value (TSV) files.

occurrences_pages :: Lens' Occurrences (Maybe [Page])

Occurrences of sensitive data in an Adobe Portable Document Format (PDF) file.

occurrences_records :: Lens' Occurrences (Maybe [Record])

Occurrences of sensitive data in an Apache Avro object container or an Apache Parquet file.

occurrences_offsetRanges :: Lens' Occurrences (Maybe [Range])

Occurrences of sensitive data detected in a binary text file.