Amazonka.SecurityHub.Types.PatchSummary

Description

Documentation

data PatchSummary

Provides an overview of the patch compliance status for an instance against a selected compliance standard.

See: newPatchSummary smart constructor.

Constructors

PatchSummary'

Fields

operationEndTime :: Maybe Text

Indicates when the operation completed.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

rebootOption :: Maybe Text

The reboot option specified for the instance.

operation :: Maybe Text

The type of patch operation performed. For Patch Manager, the values are SCAN and INSTALL.

installedRejectedCount :: Maybe Int

The number of patches that are installed but are also on a list of patches that the customer rejected.

failedCount :: Maybe Int

The number of patches from the compliance standard that failed to install.

installedOtherCount :: Maybe Int

The number of installed patches that are not part of the compliance standard.

missingCount :: Maybe Int

The number of patches that are part of the compliance standard but are not installed. The count includes patches that failed to install.

operationStartTime :: Maybe Text

Indicates when the operation started.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

installedCount :: Maybe Int

The number of patches from the compliance standard that were installed successfully.

installedPendingReboot :: Maybe Int

The number of patches that were applied, but that require the instance to be rebooted in order to be marked as installed.

id :: Text

The identifier of the compliance standard that was used to determine the patch compliance status.

newPatchSummary

Arguments

:: Text
$sel:id:PatchSummary'

-> PatchSummary

Create a value of PatchSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationEndTime:PatchSummary', patchSummary_operationEndTime - Indicates when the operation completed.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:rebootOption:PatchSummary', patchSummary_rebootOption - The reboot option specified for the instance.

$sel:operation:PatchSummary', patchSummary_operation - The type of patch operation performed. For Patch Manager, the values are SCAN and INSTALL.

$sel:installedRejectedCount:PatchSummary', patchSummary_installedRejectedCount - The number of patches that are installed but are also on a list of patches that the customer rejected.

$sel:failedCount:PatchSummary', patchSummary_failedCount - The number of patches from the compliance standard that failed to install.

$sel:installedOtherCount:PatchSummary', patchSummary_installedOtherCount - The number of installed patches that are not part of the compliance standard.

$sel:missingCount:PatchSummary', patchSummary_missingCount - The number of patches that are part of the compliance standard but are not installed. The count includes patches that failed to install.

$sel:operationStartTime:PatchSummary', patchSummary_operationStartTime - Indicates when the operation started.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

$sel:installedCount:PatchSummary', patchSummary_installedCount - The number of patches from the compliance standard that were installed successfully.

$sel:installedPendingReboot:PatchSummary', patchSummary_installedPendingReboot - The number of patches that were applied, but that require the instance to be rebooted in order to be marked as installed.

$sel:id:PatchSummary', patchSummary_id - The identifier of the compliance standard that was used to determine the patch compliance status.

patchSummary_operationEndTime :: Lens' PatchSummary (Maybe Text)

Indicates when the operation completed.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

patchSummary_rebootOption :: Lens' PatchSummary (Maybe Text)

The reboot option specified for the instance.

patchSummary_operation :: Lens' PatchSummary (Maybe Text)

The type of patch operation performed. For Patch Manager, the values are SCAN and INSTALL.

patchSummary_installedRejectedCount :: Lens' PatchSummary (Maybe Int)

The number of patches that are installed but are also on a list of patches that the customer rejected.

patchSummary_failedCount :: Lens' PatchSummary (Maybe Int)

The number of patches from the compliance standard that failed to install.

patchSummary_installedOtherCount :: Lens' PatchSummary (Maybe Int)

The number of installed patches that are not part of the compliance standard.

patchSummary_missingCount :: Lens' PatchSummary (Maybe Int)

The number of patches that are part of the compliance standard but are not installed. The count includes patches that failed to install.

patchSummary_operationStartTime :: Lens' PatchSummary (Maybe Text)

Indicates when the operation started.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

patchSummary_installedCount :: Lens' PatchSummary (Maybe Int)

The number of patches from the compliance standard that were installed successfully.

patchSummary_installedPendingReboot :: Lens' PatchSummary (Maybe Int)

The number of patches that were applied, but that require the instance to be rebooted in order to be marked as installed.

patchSummary_id :: Lens' PatchSummary Text

The identifier of the compliance standard that was used to determine the patch compliance status.