Amazonka.SecurityHub.Types.PortProbeDetail

Description

Documentation

data PortProbeDetail

A port scan that was part of the port probe. For each scan, PortProbeDetails provides information about the local IP address and port that were scanned, and the remote IP address that the scan originated from.

See: newPortProbeDetail smart constructor.

Constructors

PortProbeDetail'

Fields

remoteIpDetails :: Maybe ActionRemoteIpDetails

Provides information about the remote IP address that performed the scan.

localIpDetails :: Maybe ActionLocalIpDetails

Provides information about the IP address where the scanned port is located.

localPortDetails :: Maybe ActionLocalPortDetails

Provides information about the port that was scanned.

newPortProbeDetail :: PortProbeDetail

Create a value of PortProbeDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remoteIpDetails:PortProbeDetail', portProbeDetail_remoteIpDetails - Provides information about the remote IP address that performed the scan.

$sel:localIpDetails:PortProbeDetail', portProbeDetail_localIpDetails - Provides information about the IP address where the scanned port is located.

$sel:localPortDetails:PortProbeDetail', portProbeDetail_localPortDetails - Provides information about the port that was scanned.

portProbeDetail_remoteIpDetails :: Lens' PortProbeDetail (Maybe ActionRemoteIpDetails)

Provides information about the remote IP address that performed the scan.

portProbeDetail_localIpDetails :: Lens' PortProbeDetail (Maybe ActionLocalIpDetails)

Provides information about the IP address where the scanned port is located.

portProbeDetail_localPortDetails :: Lens' PortProbeDetail (Maybe ActionLocalPortDetails)

Provides information about the port that was scanned.