Amazonka.SecurityHub.Types.Remediation

Description

Documentation

data Remediation

Details about the remediation steps for a finding.

See: newRemediation smart constructor.

Constructors

Remediation'

Fields

recommendation :: Maybe Recommendation

A recommendation on the steps to take to remediate the issue identified by a finding.

newRemediation :: Remediation

Create a value of Remediation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommendation:Remediation', remediation_recommendation - A recommendation on the steps to take to remediate the issue identified by a finding.

remediation_recommendation :: Lens' Remediation (Maybe Recommendation)

A recommendation on the steps to take to remediate the issue identified by a finding.