Amazonka.SecurityHub.Types.Resource

Description

Documentation

data Resource

A resource related to a finding.

See: newResource smart constructor.

Constructors

Resource'

Fields

dataClassification :: Maybe DataClassificationDetails

Contains information about sensitive data that was detected on the resource.

partition :: Maybe Partition

The canonical Amazon Web Services partition name that the Region is assigned to.

details :: Maybe ResourceDetails

Additional details about the resource related to a finding.

region :: Maybe Text

The canonical Amazon Web Services external Region name where this resource is located.

resourceRole :: Maybe Text

Identifies the role of the resource in the finding. A resource is either the actor or target of the finding activity,

tags :: Maybe (HashMap Text Text)

A list of Amazon Web Services tags associated with a resource at the time the finding was processed.

type' :: Text

The type of the resource that details are provided for. If possible, set Type to one of the supported resource types. For example, if the resource is an EC2 instance, then set Type to AwsEc2Instance.

If the resource does not match any of the provided types, then set Type to Other.

id :: Text

The canonical identifier for the given resource type.

newResource

Arguments

:: Text
$sel:type':Resource'

-> Text
$sel:id:Resource'

-> Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataClassification:Resource', resource_dataClassification - Contains information about sensitive data that was detected on the resource.

$sel:partition:Resource', resource_partition - The canonical Amazon Web Services partition name that the Region is assigned to.

$sel:details:Resource', resource_details - Additional details about the resource related to a finding.

$sel:region:Resource', resource_region - The canonical Amazon Web Services external Region name where this resource is located.

$sel:resourceRole:Resource', resource_resourceRole - Identifies the role of the resource in the finding. A resource is either the actor or target of the finding activity,

$sel:tags:Resource', resource_tags - A list of Amazon Web Services tags associated with a resource at the time the finding was processed.

$sel:type':Resource', resource_type - The type of the resource that details are provided for. If possible, set Type to one of the supported resource types. For example, if the resource is an EC2 instance, then set Type to AwsEc2Instance.

If the resource does not match any of the provided types, then set Type to Other.

$sel:id:Resource', resource_id - The canonical identifier for the given resource type.

resource_dataClassification :: Lens' Resource (Maybe DataClassificationDetails)

Contains information about sensitive data that was detected on the resource.

resource_partition :: Lens' Resource (Maybe Partition)

The canonical Amazon Web Services partition name that the Region is assigned to.

resource_details :: Lens' Resource (Maybe ResourceDetails)

Additional details about the resource related to a finding.

resource_region :: Lens' Resource (Maybe Text)

The canonical Amazon Web Services external Region name where this resource is located.

resource_resourceRole :: Lens' Resource (Maybe Text)

Identifies the role of the resource in the finding. A resource is either the actor or target of the finding activity,

resource_tags :: Lens' Resource (Maybe (HashMap Text Text))

A list of Amazon Web Services tags associated with a resource at the time the finding was processed.

resource_type :: Lens' Resource Text

The type of the resource that details are provided for. If possible, set Type to one of the supported resource types. For example, if the resource is an EC2 instance, then set Type to AwsEc2Instance.

If the resource does not match any of the provided types, then set Type to Other.

resource_id :: Lens' Resource Text

The canonical identifier for the given resource type.