Amazonka.SecurityHub.Types.ResourceDetails

Description

Documentation

data ResourceDetails

Additional details about a resource related to a finding.

To provide the details, use the object that corresponds to the resource type. For example, if the resource type is AwsEc2Instance, then you use the AwsEc2Instance object to provide the details.

If the type-specific object does not contain all of the fields you want to populate, then you use the Other object to populate those additional fields.

You also use the Other object to populate the details when the selected type does not have a corresponding object.

See: newResourceDetails smart constructor.

Constructors

ResourceDetails'

Fields

awsCloudTrailTrail :: Maybe AwsCloudTrailTrailDetails

Provides details about a CloudTrail trail.

awsRdsDbCluster :: Maybe AwsRdsDbClusterDetails

Details about an Amazon RDS database cluster.

other :: Maybe (HashMap Text Text)

Details about a resource that are not available in a type-specific details object. Use the Other object in the following cases.

The type-specific object does not contain all of the fields that you want to populate. In this case, first use the type-specific object to populate those fields. Use the Other object to populate the fields that are missing from the type-specific object.

The resource type does not have a corresponding object. This includes resources for which the type is Other.

awsIamRole :: Maybe AwsIamRoleDetails

Details about an IAM role.

awsElbLoadBalancer :: Maybe AwsElbLoadBalancerDetails

Contains details about a Classic Load Balancer.

awsEc2VpnConnection :: Maybe AwsEc2VpnConnectionDetails

Details about an EC2 VPN connection.

awsCloudFrontDistribution :: Maybe AwsCloudFrontDistributionDetails

Details about a CloudFront distribution.

awsKmsKey :: Maybe AwsKmsKeyDetails

Details about an KMS key.

awsEksCluster :: Maybe AwsEksClusterDetails

Details about an Amazon EKS cluster.

awsAutoScalingLaunchConfiguration :: Maybe AwsAutoScalingLaunchConfigurationDetails

Provides details about a launch configuration.

awsOpenSearchServiceDomain :: Maybe AwsOpenSearchServiceDomainDetails

Details about an Amazon OpenSearch Service domain.

awsRdsDbSnapshot :: Maybe AwsRdsDbSnapshotDetails

Details about an Amazon RDS database snapshot.

awsWafRegionalRateBasedRule :: Maybe AwsWafRegionalRateBasedRuleDetails

Details about a rate-based rule for Regional resources.

awsApiGatewayRestApi :: Maybe AwsApiGatewayRestApiDetails

Provides information about a REST API in version 1 of Amazon API Gateway.

awsSqsQueue :: Maybe AwsSqsQueueDetails

Details about an SQS queue.

awsSecretsManagerSecret :: Maybe AwsSecretsManagerSecretDetails

Details about a Secrets Manager secret.

awsDynamoDbTable :: Maybe AwsDynamoDbTableDetails

Details about a DynamoDB table.

awsS3AccountPublicAccessBlock :: Maybe AwsS3AccountPublicAccessBlockDetails

Details about the Amazon S3 Public Access Block configuration for an account.

awsRdsDbClusterSnapshot :: Maybe AwsRdsDbClusterSnapshotDetails

Details about an Amazon RDS database cluster snapshot.

awsEc2SecurityGroup :: Maybe AwsEc2SecurityGroupDetails

Details for an EC2 security group.

awsEc2Instance :: Maybe AwsEc2InstanceDetails

Details about an EC2 instance related to a finding.

awsIamPolicy :: Maybe AwsIamPolicyDetails

Details about an IAM permissions policy.

awsS3Object :: Maybe AwsS3ObjectDetails

Details about an S3 object related to a finding.

awsEcsTaskDefinition :: Maybe AwsEcsTaskDefinitionDetails

Details about a task definition. A task definition describes the container and volume definitions of an Amazon Elastic Container Service task.

awsLambdaLayerVersion :: Maybe AwsLambdaLayerVersionDetails

Details for a Lambda layer version.

awsS3Bucket :: Maybe AwsS3BucketDetails

Details about an S3 bucket related to a finding.

awsIamAccessKey :: Maybe AwsIamAccessKeyDetails

Details about an IAM access key related to a finding.

awsCodeBuildProject :: Maybe AwsCodeBuildProjectDetails

Details for an CodeBuild project.

awsApiGatewayStage :: Maybe AwsApiGatewayStageDetails

Provides information about a version 1 Amazon API Gateway stage.

awsEc2NetworkAcl :: Maybe AwsEc2NetworkAclDetails

Details about an EC2 network access control list (ACL).

awsApiGatewayV2Api :: Maybe AwsApiGatewayV2ApiDetails

Provides information about a version 2 API in Amazon API Gateway.

awsElasticBeanstalkEnvironment :: Maybe AwsElasticBeanstalkEnvironmentDetails

Details about an Elastic Beanstalk environment.

awsEc2Vpc :: Maybe AwsEc2VpcDetails

Details for an EC2 VPC.

awsRedshiftCluster :: Maybe AwsRedshiftClusterDetails

Contains details about an Amazon Redshift cluster.

awsRdsDbInstance :: Maybe AwsRdsDbInstanceDetails

Details about an Amazon RDS database instance.

awsRdsEventSubscription :: Maybe AwsRdsEventSubscriptionDetails

Details about an RDS event notification subscription.

container :: Maybe ContainerDetails

Details about a container resource related to a finding.

awsIamGroup :: Maybe AwsIamGroupDetails

Contains details about an IAM group.

awsXrayEncryptionConfig :: Maybe AwsXrayEncryptionConfigDetails

Information about the encryption configuration for X-Ray.

awsEc2Subnet :: Maybe AwsEc2SubnetDetails

Details about a subnet in Amazon EC2.

awsEcsCluster :: Maybe AwsEcsClusterDetails

Details about an ECS cluster.

awsEc2VpcEndpointService :: Maybe AwsEc2VpcEndpointServiceDetails

Details about the service configuration for a VPC endpoint service.

awsAutoScalingAutoScalingGroup :: Maybe AwsAutoScalingAutoScalingGroupDetails

Details for an autoscaling group.

awsSnsTopic :: Maybe AwsSnsTopicDetails

Details about an SNS topic.

awsEc2NetworkInterface :: Maybe AwsEc2NetworkInterfaceDetails

Details for an EC2 network interface.

awsEcsService :: Maybe AwsEcsServiceDetails

Details about a service within an ECS cluster.

awsEc2Eip :: Maybe AwsEc2EipDetails

Details about an Elastic IP address.

awsApiGatewayV2Stage :: Maybe AwsApiGatewayV2StageDetails

Provides information about a version 2 stage for Amazon API Gateway.

awsLambdaFunction :: Maybe AwsLambdaFunctionDetails

Details about a Lambda function.

awsElbv2LoadBalancer :: Maybe AwsElbv2LoadBalancerDetails

Details about a load balancer.

awsWafRateBasedRule :: Maybe AwsWafRateBasedRuleDetails

Details about a rate-based rule for global resources.

awsEcrRepository :: Maybe AwsEcrRepositoryDetails

Information about an Amazon Elastic Container Registry repository.

awsWafWebAcl :: Maybe AwsWafWebAclDetails

Details for an WAF WebACL.

awsSsmPatchCompliance :: Maybe AwsSsmPatchComplianceDetails

Provides information about the state of a patch on an instance based on the patch baseline that was used to patch the instance.

awsElasticsearchDomain :: Maybe AwsElasticsearchDomainDetails

Details for an Elasticsearch domain.

awsCertificateManagerCertificate :: Maybe AwsCertificateManagerCertificateDetails

Provides details about an Certificate Manager certificate.

awsEcrContainerImage :: Maybe AwsEcrContainerImageDetails

Information about an Amazon ECR image.

awsIamUser :: Maybe AwsIamUserDetails

Details about an IAM user.

awsEc2Volume :: Maybe AwsEc2VolumeDetails

Details for an EC2 volume.

newResourceDetails :: ResourceDetails

Create a value of ResourceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsCloudTrailTrail:ResourceDetails', resourceDetails_awsCloudTrailTrail - Provides details about a CloudTrail trail.

$sel:awsRdsDbCluster:ResourceDetails', resourceDetails_awsRdsDbCluster - Details about an Amazon RDS database cluster.

$sel:other:ResourceDetails', resourceDetails_other - Details about a resource that are not available in a type-specific details object. Use the Other object in the following cases.

The type-specific object does not contain all of the fields that you want to populate. In this case, first use the type-specific object to populate those fields. Use the Other object to populate the fields that are missing from the type-specific object.

The resource type does not have a corresponding object. This includes resources for which the type is Other.

$sel:awsIamRole:ResourceDetails', resourceDetails_awsIamRole - Details about an IAM role.

$sel:awsElbLoadBalancer:ResourceDetails', resourceDetails_awsElbLoadBalancer - Contains details about a Classic Load Balancer.

$sel:awsEc2VpnConnection:ResourceDetails', resourceDetails_awsEc2VpnConnection - Details about an EC2 VPN connection.

$sel:awsCloudFrontDistribution:ResourceDetails', resourceDetails_awsCloudFrontDistribution - Details about a CloudFront distribution.

$sel:awsKmsKey:ResourceDetails', resourceDetails_awsKmsKey - Details about an KMS key.

$sel:awsEksCluster:ResourceDetails', resourceDetails_awsEksCluster - Details about an Amazon EKS cluster.

$sel:awsAutoScalingLaunchConfiguration:ResourceDetails', resourceDetails_awsAutoScalingLaunchConfiguration - Provides details about a launch configuration.

$sel:awsOpenSearchServiceDomain:ResourceDetails', resourceDetails_awsOpenSearchServiceDomain - Details about an Amazon OpenSearch Service domain.

$sel:awsRdsDbSnapshot:ResourceDetails', resourceDetails_awsRdsDbSnapshot - Details about an Amazon RDS database snapshot.

$sel:awsWafRegionalRateBasedRule:ResourceDetails', resourceDetails_awsWafRegionalRateBasedRule - Details about a rate-based rule for Regional resources.

$sel:awsApiGatewayRestApi:ResourceDetails', resourceDetails_awsApiGatewayRestApi - Provides information about a REST API in version 1 of Amazon API Gateway.

$sel:awsSqsQueue:ResourceDetails', resourceDetails_awsSqsQueue - Details about an SQS queue.

$sel:awsSecretsManagerSecret:ResourceDetails', resourceDetails_awsSecretsManagerSecret - Details about a Secrets Manager secret.

$sel:awsDynamoDbTable:ResourceDetails', resourceDetails_awsDynamoDbTable - Details about a DynamoDB table.

$sel:awsS3AccountPublicAccessBlock:ResourceDetails', resourceDetails_awsS3AccountPublicAccessBlock - Details about the Amazon S3 Public Access Block configuration for an account.

$sel:awsRdsDbClusterSnapshot:ResourceDetails', resourceDetails_awsRdsDbClusterSnapshot - Details about an Amazon RDS database cluster snapshot.

$sel:awsEc2SecurityGroup:ResourceDetails', resourceDetails_awsEc2SecurityGroup - Details for an EC2 security group.

$sel:awsEc2Instance:ResourceDetails', resourceDetails_awsEc2Instance - Details about an EC2 instance related to a finding.

$sel:awsIamPolicy:ResourceDetails', resourceDetails_awsIamPolicy - Details about an IAM permissions policy.

$sel:awsS3Object:ResourceDetails', resourceDetails_awsS3Object - Details about an S3 object related to a finding.

$sel:awsEcsTaskDefinition:ResourceDetails', resourceDetails_awsEcsTaskDefinition - Details about a task definition. A task definition describes the container and volume definitions of an Amazon Elastic Container Service task.

$sel:awsLambdaLayerVersion:ResourceDetails', resourceDetails_awsLambdaLayerVersion - Details for a Lambda layer version.

$sel:awsS3Bucket:ResourceDetails', resourceDetails_awsS3Bucket - Details about an S3 bucket related to a finding.

$sel:awsIamAccessKey:ResourceDetails', resourceDetails_awsIamAccessKey - Details about an IAM access key related to a finding.

$sel:awsCodeBuildProject:ResourceDetails', resourceDetails_awsCodeBuildProject - Details for an CodeBuild project.

$sel:awsApiGatewayStage:ResourceDetails', resourceDetails_awsApiGatewayStage - Provides information about a version 1 Amazon API Gateway stage.

$sel:awsEc2NetworkAcl:ResourceDetails', resourceDetails_awsEc2NetworkAcl - Details about an EC2 network access control list (ACL).

$sel:awsApiGatewayV2Api:ResourceDetails', resourceDetails_awsApiGatewayV2Api - Provides information about a version 2 API in Amazon API Gateway.

$sel:awsElasticBeanstalkEnvironment:ResourceDetails', resourceDetails_awsElasticBeanstalkEnvironment - Details about an Elastic Beanstalk environment.

$sel:awsEc2Vpc:ResourceDetails', resourceDetails_awsEc2Vpc - Details for an EC2 VPC.

$sel:awsRedshiftCluster:ResourceDetails', resourceDetails_awsRedshiftCluster - Contains details about an Amazon Redshift cluster.

$sel:awsRdsDbInstance:ResourceDetails', resourceDetails_awsRdsDbInstance - Details about an Amazon RDS database instance.

$sel:awsRdsEventSubscription:ResourceDetails', resourceDetails_awsRdsEventSubscription - Details about an RDS event notification subscription.

$sel:container:ResourceDetails', resourceDetails_container - Details about a container resource related to a finding.

$sel:awsIamGroup:ResourceDetails', resourceDetails_awsIamGroup - Contains details about an IAM group.

$sel:awsXrayEncryptionConfig:ResourceDetails', resourceDetails_awsXrayEncryptionConfig - Information about the encryption configuration for X-Ray.

$sel:awsEc2Subnet:ResourceDetails', resourceDetails_awsEc2Subnet - Details about a subnet in Amazon EC2.

$sel:awsEcsCluster:ResourceDetails', resourceDetails_awsEcsCluster - Details about an ECS cluster.

$sel:awsEc2VpcEndpointService:ResourceDetails', resourceDetails_awsEc2VpcEndpointService - Details about the service configuration for a VPC endpoint service.

$sel:awsAutoScalingAutoScalingGroup:ResourceDetails', resourceDetails_awsAutoScalingAutoScalingGroup - Details for an autoscaling group.

$sel:awsSnsTopic:ResourceDetails', resourceDetails_awsSnsTopic - Details about an SNS topic.

$sel:awsEc2NetworkInterface:ResourceDetails', resourceDetails_awsEc2NetworkInterface - Details for an EC2 network interface.

$sel:awsEcsService:ResourceDetails', resourceDetails_awsEcsService - Details about a service within an ECS cluster.

$sel:awsEc2Eip:ResourceDetails', resourceDetails_awsEc2Eip - Details about an Elastic IP address.

$sel:awsApiGatewayV2Stage:ResourceDetails', resourceDetails_awsApiGatewayV2Stage - Provides information about a version 2 stage for Amazon API Gateway.

$sel:awsLambdaFunction:ResourceDetails', resourceDetails_awsLambdaFunction - Details about a Lambda function.

$sel:awsElbv2LoadBalancer:ResourceDetails', resourceDetails_awsElbv2LoadBalancer - Details about a load balancer.

$sel:awsWafRateBasedRule:ResourceDetails', resourceDetails_awsWafRateBasedRule - Details about a rate-based rule for global resources.

$sel:awsEcrRepository:ResourceDetails', resourceDetails_awsEcrRepository - Information about an Amazon Elastic Container Registry repository.

$sel:awsWafWebAcl:ResourceDetails', resourceDetails_awsWafWebAcl - Details for an WAF WebACL.

$sel:awsSsmPatchCompliance:ResourceDetails', resourceDetails_awsSsmPatchCompliance - Provides information about the state of a patch on an instance based on the patch baseline that was used to patch the instance.

$sel:awsElasticsearchDomain:ResourceDetails', resourceDetails_awsElasticsearchDomain - Details for an Elasticsearch domain.

$sel:awsCertificateManagerCertificate:ResourceDetails', resourceDetails_awsCertificateManagerCertificate - Provides details about an Certificate Manager certificate.

$sel:awsEcrContainerImage:ResourceDetails', resourceDetails_awsEcrContainerImage - Information about an Amazon ECR image.

$sel:awsIamUser:ResourceDetails', resourceDetails_awsIamUser - Details about an IAM user.

$sel:awsEc2Volume:ResourceDetails', resourceDetails_awsEc2Volume - Details for an EC2 volume.

resourceDetails_awsCloudTrailTrail :: Lens' ResourceDetails (Maybe AwsCloudTrailTrailDetails)

Provides details about a CloudTrail trail.

resourceDetails_awsRdsDbCluster :: Lens' ResourceDetails (Maybe AwsRdsDbClusterDetails)

Details about an Amazon RDS database cluster.

resourceDetails_other :: Lens' ResourceDetails (Maybe (HashMap Text Text))

Details about a resource that are not available in a type-specific details object. Use the Other object in the following cases.

The type-specific object does not contain all of the fields that you want to populate. In this case, first use the type-specific object to populate those fields. Use the Other object to populate the fields that are missing from the type-specific object.

The resource type does not have a corresponding object. This includes resources for which the type is Other.

resourceDetails_awsIamRole :: Lens' ResourceDetails (Maybe AwsIamRoleDetails)

Details about an IAM role.

resourceDetails_awsElbLoadBalancer :: Lens' ResourceDetails (Maybe AwsElbLoadBalancerDetails)

Contains details about a Classic Load Balancer.

resourceDetails_awsEc2VpnConnection :: Lens' ResourceDetails (Maybe AwsEc2VpnConnectionDetails)

Details about an EC2 VPN connection.

resourceDetails_awsCloudFrontDistribution :: Lens' ResourceDetails (Maybe AwsCloudFrontDistributionDetails)

Details about a CloudFront distribution.

resourceDetails_awsKmsKey :: Lens' ResourceDetails (Maybe AwsKmsKeyDetails)

Details about an KMS key.

resourceDetails_awsEksCluster :: Lens' ResourceDetails (Maybe AwsEksClusterDetails)

Details about an Amazon EKS cluster.

resourceDetails_awsAutoScalingLaunchConfiguration :: Lens' ResourceDetails (Maybe AwsAutoScalingLaunchConfigurationDetails)

Provides details about a launch configuration.

resourceDetails_awsOpenSearchServiceDomain :: Lens' ResourceDetails (Maybe AwsOpenSearchServiceDomainDetails)

Details about an Amazon OpenSearch Service domain.

resourceDetails_awsRdsDbSnapshot :: Lens' ResourceDetails (Maybe AwsRdsDbSnapshotDetails)

Details about an Amazon RDS database snapshot.

resourceDetails_awsWafRegionalRateBasedRule :: Lens' ResourceDetails (Maybe AwsWafRegionalRateBasedRuleDetails)

Details about a rate-based rule for Regional resources.

resourceDetails_awsApiGatewayRestApi :: Lens' ResourceDetails (Maybe AwsApiGatewayRestApiDetails)

Provides information about a REST API in version 1 of Amazon API Gateway.

resourceDetails_awsSqsQueue :: Lens' ResourceDetails (Maybe AwsSqsQueueDetails)

Details about an SQS queue.

resourceDetails_awsSecretsManagerSecret :: Lens' ResourceDetails (Maybe AwsSecretsManagerSecretDetails)

Details about a Secrets Manager secret.

resourceDetails_awsDynamoDbTable :: Lens' ResourceDetails (Maybe AwsDynamoDbTableDetails)

Details about a DynamoDB table.

resourceDetails_awsS3AccountPublicAccessBlock :: Lens' ResourceDetails (Maybe AwsS3AccountPublicAccessBlockDetails)

Details about the Amazon S3 Public Access Block configuration for an account.

resourceDetails_awsRdsDbClusterSnapshot :: Lens' ResourceDetails (Maybe AwsRdsDbClusterSnapshotDetails)

Details about an Amazon RDS database cluster snapshot.

resourceDetails_awsEc2SecurityGroup :: Lens' ResourceDetails (Maybe AwsEc2SecurityGroupDetails)

Details for an EC2 security group.

resourceDetails_awsEc2Instance :: Lens' ResourceDetails (Maybe AwsEc2InstanceDetails)

Details about an EC2 instance related to a finding.

resourceDetails_awsIamPolicy :: Lens' ResourceDetails (Maybe AwsIamPolicyDetails)

Details about an IAM permissions policy.

resourceDetails_awsS3Object :: Lens' ResourceDetails (Maybe AwsS3ObjectDetails)

Details about an S3 object related to a finding.

resourceDetails_awsEcsTaskDefinition :: Lens' ResourceDetails (Maybe AwsEcsTaskDefinitionDetails)

Details about a task definition. A task definition describes the container and volume definitions of an Amazon Elastic Container Service task.

resourceDetails_awsLambdaLayerVersion :: Lens' ResourceDetails (Maybe AwsLambdaLayerVersionDetails)

Details for a Lambda layer version.

resourceDetails_awsS3Bucket :: Lens' ResourceDetails (Maybe AwsS3BucketDetails)

Details about an S3 bucket related to a finding.

resourceDetails_awsIamAccessKey :: Lens' ResourceDetails (Maybe AwsIamAccessKeyDetails)

Details about an IAM access key related to a finding.

resourceDetails_awsCodeBuildProject :: Lens' ResourceDetails (Maybe AwsCodeBuildProjectDetails)

Details for an CodeBuild project.

resourceDetails_awsApiGatewayStage :: Lens' ResourceDetails (Maybe AwsApiGatewayStageDetails)

Provides information about a version 1 Amazon API Gateway stage.

resourceDetails_awsEc2NetworkAcl :: Lens' ResourceDetails (Maybe AwsEc2NetworkAclDetails)

Details about an EC2 network access control list (ACL).

resourceDetails_awsApiGatewayV2Api :: Lens' ResourceDetails (Maybe AwsApiGatewayV2ApiDetails)

Provides information about a version 2 API in Amazon API Gateway.

resourceDetails_awsElasticBeanstalkEnvironment :: Lens' ResourceDetails (Maybe AwsElasticBeanstalkEnvironmentDetails)

Details about an Elastic Beanstalk environment.

resourceDetails_awsEc2Vpc :: Lens' ResourceDetails (Maybe AwsEc2VpcDetails)

Details for an EC2 VPC.

resourceDetails_awsRedshiftCluster :: Lens' ResourceDetails (Maybe AwsRedshiftClusterDetails)

Contains details about an Amazon Redshift cluster.

resourceDetails_awsRdsDbInstance :: Lens' ResourceDetails (Maybe AwsRdsDbInstanceDetails)

Details about an Amazon RDS database instance.

resourceDetails_awsRdsEventSubscription :: Lens' ResourceDetails (Maybe AwsRdsEventSubscriptionDetails)

Details about an RDS event notification subscription.

resourceDetails_container :: Lens' ResourceDetails (Maybe ContainerDetails)

Details about a container resource related to a finding.

resourceDetails_awsIamGroup :: Lens' ResourceDetails (Maybe AwsIamGroupDetails)

Contains details about an IAM group.

resourceDetails_awsXrayEncryptionConfig :: Lens' ResourceDetails (Maybe AwsXrayEncryptionConfigDetails)

Information about the encryption configuration for X-Ray.

resourceDetails_awsEc2Subnet :: Lens' ResourceDetails (Maybe AwsEc2SubnetDetails)

Details about a subnet in Amazon EC2.

resourceDetails_awsEcsCluster :: Lens' ResourceDetails (Maybe AwsEcsClusterDetails)

Details about an ECS cluster.

resourceDetails_awsEc2VpcEndpointService :: Lens' ResourceDetails (Maybe AwsEc2VpcEndpointServiceDetails)

Details about the service configuration for a VPC endpoint service.

resourceDetails_awsAutoScalingAutoScalingGroup :: Lens' ResourceDetails (Maybe AwsAutoScalingAutoScalingGroupDetails)

Details for an autoscaling group.

resourceDetails_awsSnsTopic :: Lens' ResourceDetails (Maybe AwsSnsTopicDetails)

Details about an SNS topic.

resourceDetails_awsEc2NetworkInterface :: Lens' ResourceDetails (Maybe AwsEc2NetworkInterfaceDetails)

Details for an EC2 network interface.

resourceDetails_awsEcsService :: Lens' ResourceDetails (Maybe AwsEcsServiceDetails)

Details about a service within an ECS cluster.

resourceDetails_awsEc2Eip :: Lens' ResourceDetails (Maybe AwsEc2EipDetails)

Details about an Elastic IP address.

resourceDetails_awsApiGatewayV2Stage :: Lens' ResourceDetails (Maybe AwsApiGatewayV2StageDetails)

Provides information about a version 2 stage for Amazon API Gateway.

resourceDetails_awsLambdaFunction :: Lens' ResourceDetails (Maybe AwsLambdaFunctionDetails)

Details about a Lambda function.

resourceDetails_awsElbv2LoadBalancer :: Lens' ResourceDetails (Maybe AwsElbv2LoadBalancerDetails)

Details about a load balancer.

resourceDetails_awsWafRateBasedRule :: Lens' ResourceDetails (Maybe AwsWafRateBasedRuleDetails)

Details about a rate-based rule for global resources.

resourceDetails_awsEcrRepository :: Lens' ResourceDetails (Maybe AwsEcrRepositoryDetails)

Information about an Amazon Elastic Container Registry repository.

resourceDetails_awsWafWebAcl :: Lens' ResourceDetails (Maybe AwsWafWebAclDetails)

Details for an WAF WebACL.

resourceDetails_awsSsmPatchCompliance :: Lens' ResourceDetails (Maybe AwsSsmPatchComplianceDetails)

Provides information about the state of a patch on an instance based on the patch baseline that was used to patch the instance.

resourceDetails_awsElasticsearchDomain :: Lens' ResourceDetails (Maybe AwsElasticsearchDomainDetails)

Details for an Elasticsearch domain.

resourceDetails_awsCertificateManagerCertificate :: Lens' ResourceDetails (Maybe AwsCertificateManagerCertificateDetails)

Provides details about an Certificate Manager certificate.

resourceDetails_awsEcrContainerImage :: Lens' ResourceDetails (Maybe AwsEcrContainerImageDetails)

Information about an Amazon ECR image.

resourceDetails_awsIamUser :: Lens' ResourceDetails (Maybe AwsIamUserDetails)

Details about an IAM user.

resourceDetails_awsEc2Volume :: Lens' ResourceDetails (Maybe AwsEc2VolumeDetails)

Details for an EC2 volume.