Amazonka.SecurityHub.Types.SensitiveDataResult

Description

Documentation

data SensitiveDataResult

Contains a detected instance of sensitive data that are based on built-in identifiers.

See: newSensitiveDataResult smart constructor.

Constructors

SensitiveDataResult'

Fields

detections :: Maybe [SensitiveDataDetections]

The list of detected instances of sensitive data.

category :: Maybe Text

The category of sensitive data that was detected. For example, the category can indicate that the sensitive data involved credentials, financial information, or personal information.

totalCount :: Maybe Integer

The total number of occurrences of sensitive data.

newSensitiveDataResult :: SensitiveDataResult

Create a value of SensitiveDataResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detections:SensitiveDataResult', sensitiveDataResult_detections - The list of detected instances of sensitive data.

$sel:category:SensitiveDataResult', sensitiveDataResult_category - The category of sensitive data that was detected. For example, the category can indicate that the sensitive data involved credentials, financial information, or personal information.

$sel:totalCount:SensitiveDataResult', sensitiveDataResult_totalCount - The total number of occurrences of sensitive data.

sensitiveDataResult_detections :: Lens' SensitiveDataResult (Maybe [SensitiveDataDetections])

The list of detected instances of sensitive data.

sensitiveDataResult_category :: Lens' SensitiveDataResult (Maybe Text)

The category of sensitive data that was detected. For example, the category can indicate that the sensitive data involved credentials, financial information, or personal information.

sensitiveDataResult_totalCount :: Lens' SensitiveDataResult (Maybe Integer)

The total number of occurrences of sensitive data.