Amazonka.SecurityHub.Types.SeverityUpdate

Description

Documentation

data SeverityUpdate

Updates to the severity information for a finding.

See: newSeverityUpdate smart constructor.

Constructors

SeverityUpdate'

Fields

product :: Maybe Double

The native severity as defined by the Amazon Web Services service or integrated partner product that generated the finding.

label :: Maybe SeverityLabel

The severity value of the finding. The allowed values are the following.

INFORMATIONAL - No issue was found.

LOW - The issue does not require action on its own.

MEDIUM - The issue must be addressed but not urgently.

HIGH - The issue must be addressed as a priority.

CRITICAL - The issue must be remediated immediately to avoid it escalating.

normalized :: Maybe Natural

The normalized severity for the finding. This attribute is to be deprecated in favor of Label.

If you provide Normalized and do not provide Label, Label is set automatically as follows.

0 - INFORMATIONAL

1–39 - LOW

40–69 - MEDIUM

70–89 - HIGH

90–100 - CRITICAL

newSeverityUpdate :: SeverityUpdate

Create a value of SeverityUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:product:SeverityUpdate', severityUpdate_product - The native severity as defined by the Amazon Web Services service or integrated partner product that generated the finding.

$sel:label:SeverityUpdate', severityUpdate_label - The severity value of the finding. The allowed values are the following.

INFORMATIONAL - No issue was found.

LOW - The issue does not require action on its own.

MEDIUM - The issue must be addressed but not urgently.

HIGH - The issue must be addressed as a priority.

CRITICAL - The issue must be remediated immediately to avoid it escalating.

$sel:normalized:SeverityUpdate', severityUpdate_normalized - The normalized severity for the finding. This attribute is to be deprecated in favor of Label.

If you provide Normalized and do not provide Label, Label is set automatically as follows.

0 - INFORMATIONAL

1–39 - LOW

40–69 - MEDIUM

70–89 - HIGH

90–100 - CRITICAL

severityUpdate_product :: Lens' SeverityUpdate (Maybe Double)

The native severity as defined by the Amazon Web Services service or integrated partner product that generated the finding.

severityUpdate_label :: Lens' SeverityUpdate (Maybe SeverityLabel)

The severity value of the finding. The allowed values are the following.

INFORMATIONAL - No issue was found.

LOW - The issue does not require action on its own.

MEDIUM - The issue must be addressed but not urgently.

HIGH - The issue must be addressed as a priority.

CRITICAL - The issue must be remediated immediately to avoid it escalating.

severityUpdate_normalized :: Lens' SeverityUpdate (Maybe Natural)

The normalized severity for the finding. This attribute is to be deprecated in favor of Label.

If you provide Normalized and do not provide Label, Label is set automatically as follows.

0 - INFORMATIONAL

1–39 - LOW

40–69 - MEDIUM

70–89 - HIGH

90–100 - CRITICAL