Amazonka.SecurityHub.Types.StandardsControl

Description

Documentation

data StandardsControl

Details for an individual security standard control.

See: newStandardsControl smart constructor.

Constructors

StandardsControl'

Fields

remediationUrl :: Maybe Text

A link to remediation information for the control in the Security Hub user documentation.

severityRating :: Maybe SeverityRating

The severity of findings generated from this security standard control.

The finding severity is based on an assessment of how easy it would be to compromise Amazon Web Services resources if the issue is detected.

controlStatusUpdatedAt :: Maybe POSIX

The date and time that the status of the security standard control was most recently updated.

relatedRequirements :: Maybe [Text]

The list of requirements that are related to this control.

controlStatus :: Maybe ControlStatus

The current status of the security standard control. Indicates whether the control is enabled or disabled. Security Hub does not check against disabled controls.

disabledReason :: Maybe Text

The reason provided for the most recent change in status for the control.

controlId :: Maybe Text

The identifier of the security standard control.

standardsControlArn :: Maybe Text

The ARN of the security standard control.

title :: Maybe Text

The title of the security standard control.

description :: Maybe Text

The longer description of the security standard control. Provides information about what the control is checking for.

newStandardsControl :: StandardsControl

Create a value of StandardsControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:remediationUrl:StandardsControl', standardsControl_remediationUrl - A link to remediation information for the control in the Security Hub user documentation.

$sel:severityRating:StandardsControl', standardsControl_severityRating - The severity of findings generated from this security standard control.

The finding severity is based on an assessment of how easy it would be to compromise Amazon Web Services resources if the issue is detected.

$sel:controlStatusUpdatedAt:StandardsControl', standardsControl_controlStatusUpdatedAt - The date and time that the status of the security standard control was most recently updated.

$sel:relatedRequirements:StandardsControl', standardsControl_relatedRequirements - The list of requirements that are related to this control.

$sel:controlStatus:StandardsControl', standardsControl_controlStatus - The current status of the security standard control. Indicates whether the control is enabled or disabled. Security Hub does not check against disabled controls.

$sel:disabledReason:StandardsControl', standardsControl_disabledReason - The reason provided for the most recent change in status for the control.

$sel:controlId:StandardsControl', standardsControl_controlId - The identifier of the security standard control.

$sel:standardsControlArn:StandardsControl', standardsControl_standardsControlArn - The ARN of the security standard control.

$sel:title:StandardsControl', standardsControl_title - The title of the security standard control.

$sel:description:StandardsControl', standardsControl_description - The longer description of the security standard control. Provides information about what the control is checking for.

standardsControl_remediationUrl :: Lens' StandardsControl (Maybe Text)

A link to remediation information for the control in the Security Hub user documentation.

standardsControl_severityRating :: Lens' StandardsControl (Maybe SeverityRating)

The severity of findings generated from this security standard control.

The finding severity is based on an assessment of how easy it would be to compromise Amazon Web Services resources if the issue is detected.

standardsControl_controlStatusUpdatedAt :: Lens' StandardsControl (Maybe UTCTime)

The date and time that the status of the security standard control was most recently updated.

standardsControl_relatedRequirements :: Lens' StandardsControl (Maybe [Text])

The list of requirements that are related to this control.

standardsControl_controlStatus :: Lens' StandardsControl (Maybe ControlStatus)

The current status of the security standard control. Indicates whether the control is enabled or disabled. Security Hub does not check against disabled controls.

standardsControl_disabledReason :: Lens' StandardsControl (Maybe Text)

The reason provided for the most recent change in status for the control.

standardsControl_controlId :: Lens' StandardsControl (Maybe Text)

The identifier of the security standard control.

standardsControl_standardsControlArn :: Lens' StandardsControl (Maybe Text)

The ARN of the security standard control.

standardsControl_title :: Lens' StandardsControl (Maybe Text)

The title of the security standard control.

standardsControl_description :: Lens' StandardsControl (Maybe Text)

The longer description of the security standard control. Provides information about what the control is checking for.