Amazonka.SecurityHub.Types.StandardsSubscription

Description

Documentation

data StandardsSubscription

A resource that represents your subscription to a supported standard.

See: newStandardsSubscription smart constructor.

Constructors

StandardsSubscription'

Fields

standardsSubscriptionArn :: Text

The ARN of a resource that represents your subscription to a supported standard.

standardsArn :: Text

The ARN of a standard.

standardsInput :: HashMap Text Text

A key-value pair of input for the standard.

standardsStatus :: StandardsStatus

The status of the standard subscription.

The status values are as follows:

PENDING - Standard is in the process of being enabled.

READY - Standard is enabled.

INCOMPLETE - Standard could not be enabled completely. Some controls may not be available.

DELETING - Standard is in the process of being disabled.

FAILED - Standard could not be disabled.

newStandardsSubscription

Arguments

:: Text
$sel:standardsSubscriptionArn:StandardsSubscription'

-> Text
$sel:standardsArn:StandardsSubscription'

-> StandardsStatus
$sel:standardsStatus:StandardsSubscription'

-> StandardsSubscription

Create a value of StandardsSubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:standardsSubscriptionArn:StandardsSubscription', standardsSubscription_standardsSubscriptionArn - The ARN of a resource that represents your subscription to a supported standard.

$sel:standardsArn:StandardsSubscription', standardsSubscription_standardsArn - The ARN of a standard.

$sel:standardsInput:StandardsSubscription', standardsSubscription_standardsInput - A key-value pair of input for the standard.

$sel:standardsStatus:StandardsSubscription', standardsSubscription_standardsStatus - The status of the standard subscription.

The status values are as follows:

PENDING - Standard is in the process of being enabled.

READY - Standard is enabled.

INCOMPLETE - Standard could not be enabled completely. Some controls may not be available.

DELETING - Standard is in the process of being disabled.

FAILED - Standard could not be disabled.

standardsSubscription_standardsSubscriptionArn :: Lens' StandardsSubscription Text

The ARN of a resource that represents your subscription to a supported standard.

standardsSubscription_standardsArn :: Lens' StandardsSubscription Text

The ARN of a standard.

standardsSubscription_standardsInput :: Lens' StandardsSubscription (HashMap Text Text)

A key-value pair of input for the standard.

standardsSubscription_standardsStatus :: Lens' StandardsSubscription StandardsStatus

The status of the standard subscription.

The status values are as follows:

PENDING - Standard is in the process of being enabled.

READY - Standard is enabled.

INCOMPLETE - Standard could not be enabled completely. Some controls may not be available.

DELETING - Standard is in the process of being disabled.

FAILED - Standard could not be disabled.