Amazonka.SecurityHub.UpdateFindingAggregator

Description

Updates the finding aggregation configuration. Used to update the Region linking mode and the list of included or excluded Regions. You cannot use UpdateFindingAggregator to change the aggregation Region.

You must run UpdateFindingAggregator from the current aggregation Region.

Creating a Request

data UpdateFindingAggregator

See: newUpdateFindingAggregator smart constructor.

Constructors

UpdateFindingAggregator'

Fields

regions :: Maybe [Text]

If RegionLinkingMode is ALL_REGIONS_EXCEPT_SPECIFIED, then this is a comma-separated list of Regions that do not aggregate findings to the aggregation Region.

If RegionLinkingMode is SPECIFIED_REGIONS, then this is a comma-separated list of Regions that do aggregate findings to the aggregation Region.

findingAggregatorArn :: Text

The ARN of the finding aggregator. To obtain the ARN, use ListFindingAggregators.

regionLinkingMode :: Text

Indicates whether to aggregate findings from all of the available Regions in the current partition. Also determines whether to automatically aggregate findings from new Regions as Security Hub supports them and you opt into them.

The selected option also determines how to use the Regions provided in the Regions list.

The options are as follows:

ALL_REGIONS - Indicates to aggregate findings from all of the Regions where Security Hub is enabled. When you choose this option, Security Hub also automatically aggregates findings from new Regions as Security Hub supports them and you opt into them.

ALL_REGIONS_EXCEPT_SPECIFIED - Indicates to aggregate findings from all of the Regions where Security Hub is enabled, except for the Regions listed in the Regions parameter. When you choose this option, Security Hub also automatically aggregates findings from new Regions as Security Hub supports them and you opt into them.

SPECIFIED_REGIONS - Indicates to aggregate findings only from the Regions listed in the Regions parameter. Security Hub does not automatically aggregate findings from new Regions.

newUpdateFindingAggregator

Arguments

:: Text
$sel:findingAggregatorArn:UpdateFindingAggregator'

-> Text
$sel:regionLinkingMode:UpdateFindingAggregator'

-> UpdateFindingAggregator

Create a value of UpdateFindingAggregator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regions:UpdateFindingAggregator', updateFindingAggregator_regions - If RegionLinkingMode is ALL_REGIONS_EXCEPT_SPECIFIED, then this is a comma-separated list of Regions that do not aggregate findings to the aggregation Region.

If RegionLinkingMode is SPECIFIED_REGIONS, then this is a comma-separated list of Regions that do aggregate findings to the aggregation Region.

$sel:findingAggregatorArn:UpdateFindingAggregator', updateFindingAggregator_findingAggregatorArn - The ARN of the finding aggregator. To obtain the ARN, use ListFindingAggregators.

$sel:regionLinkingMode:UpdateFindingAggregator', updateFindingAggregator_regionLinkingMode - Indicates whether to aggregate findings from all of the available Regions in the current partition. Also determines whether to automatically aggregate findings from new Regions as Security Hub supports them and you opt into them.

The selected option also determines how to use the Regions provided in the Regions list.

The options are as follows:

ALL_REGIONS - Indicates to aggregate findings from all of the Regions where Security Hub is enabled. When you choose this option, Security Hub also automatically aggregates findings from new Regions as Security Hub supports them and you opt into them.

ALL_REGIONS_EXCEPT_SPECIFIED - Indicates to aggregate findings from all of the Regions where Security Hub is enabled, except for the Regions listed in the Regions parameter. When you choose this option, Security Hub also automatically aggregates findings from new Regions as Security Hub supports them and you opt into them.

SPECIFIED_REGIONS - Indicates to aggregate findings only from the Regions listed in the Regions parameter. Security Hub does not automatically aggregate findings from new Regions.

Request Lenses

updateFindingAggregator_regions :: Lens' UpdateFindingAggregator (Maybe [Text])

If RegionLinkingMode is ALL_REGIONS_EXCEPT_SPECIFIED, then this is a comma-separated list of Regions that do not aggregate findings to the aggregation Region.

If RegionLinkingMode is SPECIFIED_REGIONS, then this is a comma-separated list of Regions that do aggregate findings to the aggregation Region.

updateFindingAggregator_findingAggregatorArn :: Lens' UpdateFindingAggregator Text

The ARN of the finding aggregator. To obtain the ARN, use ListFindingAggregators.

updateFindingAggregator_regionLinkingMode :: Lens' UpdateFindingAggregator Text

Indicates whether to aggregate findings from all of the available Regions in the current partition. Also determines whether to automatically aggregate findings from new Regions as Security Hub supports them and you opt into them.

The selected option also determines how to use the Regions provided in the Regions list.

The options are as follows:

ALL_REGIONS - Indicates to aggregate findings from all of the Regions where Security Hub is enabled. When you choose this option, Security Hub also automatically aggregates findings from new Regions as Security Hub supports them and you opt into them.

ALL_REGIONS_EXCEPT_SPECIFIED - Indicates to aggregate findings from all of the Regions where Security Hub is enabled, except for the Regions listed in the Regions parameter. When you choose this option, Security Hub also automatically aggregates findings from new Regions as Security Hub supports them and you opt into them.

SPECIFIED_REGIONS - Indicates to aggregate findings only from the Regions listed in the Regions parameter. Security Hub does not automatically aggregate findings from new Regions.

Destructuring the Response

data UpdateFindingAggregatorResponse

See: newUpdateFindingAggregatorResponse smart constructor.

Constructors

UpdateFindingAggregatorResponse'

Fields

regions :: Maybe [Text]

The list of excluded Regions or included Regions.

findingAggregatorArn :: Maybe Text

The ARN of the finding aggregator.

regionLinkingMode :: Maybe Text

Indicates whether to link all Regions, all Regions except for a list of excluded Regions, or a list of included Regions.

findingAggregationRegion :: Maybe Text

The aggregation Region.

httpStatus :: Int

The response's http status code.

newUpdateFindingAggregatorResponse

Arguments

:: Int
$sel:httpStatus:UpdateFindingAggregatorResponse'

-> UpdateFindingAggregatorResponse

Create a value of UpdateFindingAggregatorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regions:UpdateFindingAggregator', updateFindingAggregatorResponse_regions - The list of excluded Regions or included Regions.

$sel:findingAggregatorArn:UpdateFindingAggregator', updateFindingAggregatorResponse_findingAggregatorArn - The ARN of the finding aggregator.

$sel:regionLinkingMode:UpdateFindingAggregator', updateFindingAggregatorResponse_regionLinkingMode - Indicates whether to link all Regions, all Regions except for a list of excluded Regions, or a list of included Regions.

$sel:findingAggregationRegion:UpdateFindingAggregatorResponse', updateFindingAggregatorResponse_findingAggregationRegion - The aggregation Region.

$sel:httpStatus:UpdateFindingAggregatorResponse', updateFindingAggregatorResponse_httpStatus - The response's http status code.

Response Lenses

updateFindingAggregatorResponse_regions :: Lens' UpdateFindingAggregatorResponse (Maybe [Text])

The list of excluded Regions or included Regions.

updateFindingAggregatorResponse_findingAggregatorArn :: Lens' UpdateFindingAggregatorResponse (Maybe Text)

The ARN of the finding aggregator.

updateFindingAggregatorResponse_regionLinkingMode :: Lens' UpdateFindingAggregatorResponse (Maybe Text)

Indicates whether to link all Regions, all Regions except for a list of excluded Regions, or a list of included Regions.

updateFindingAggregatorResponse_findingAggregationRegion :: Lens' UpdateFindingAggregatorResponse (Maybe Text)

The aggregation Region.

updateFindingAggregatorResponse_httpStatus :: Lens' UpdateFindingAggregatorResponse Int

The response's http status code.