Amazonka.ServerlessApplicationRepository.CreateApplicationVersion

Description

Creates an application version.

Creating a Request

data CreateApplicationVersion

See: newCreateApplicationVersion smart constructor.

Constructors

CreateApplicationVersion'

Fields

sourceCodeUrl :: Maybe Text

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

templateBody :: Maybe Text

The raw packaged AWS SAM template of your application.

templateUrl :: Maybe Text

A link to the packaged AWS SAM template of your application.

sourceCodeArchiveUrl :: Maybe Text

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

applicationId :: Text

The Amazon Resource Name (ARN) of the application.

semanticVersion :: Text

The semantic version of the new version.

newCreateApplicationVersion

Arguments

:: Text
$sel:applicationId:CreateApplicationVersion'

-> Text
$sel:semanticVersion:CreateApplicationVersion'

-> CreateApplicationVersion

Create a value of CreateApplicationVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceCodeUrl:CreateApplicationVersion', createApplicationVersion_sourceCodeUrl - A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

$sel:templateBody:CreateApplicationVersion', createApplicationVersion_templateBody - The raw packaged AWS SAM template of your application.

$sel:templateUrl:CreateApplicationVersion', createApplicationVersion_templateUrl - A link to the packaged AWS SAM template of your application.

$sel:sourceCodeArchiveUrl:CreateApplicationVersion', createApplicationVersion_sourceCodeArchiveUrl - A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

$sel:applicationId:CreateApplicationVersion', createApplicationVersion_applicationId - The Amazon Resource Name (ARN) of the application.

$sel:semanticVersion:CreateApplicationVersion', createApplicationVersion_semanticVersion - The semantic version of the new version.

Request Lenses

createApplicationVersion_sourceCodeUrl :: Lens' CreateApplicationVersion (Maybe Text)

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

createApplicationVersion_templateBody :: Lens' CreateApplicationVersion (Maybe Text)

The raw packaged AWS SAM template of your application.

createApplicationVersion_templateUrl :: Lens' CreateApplicationVersion (Maybe Text)

A link to the packaged AWS SAM template of your application.

createApplicationVersion_sourceCodeArchiveUrl :: Lens' CreateApplicationVersion (Maybe Text)

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

createApplicationVersion_applicationId :: Lens' CreateApplicationVersion Text

The Amazon Resource Name (ARN) of the application.

createApplicationVersion_semanticVersion :: Lens' CreateApplicationVersion Text

The semantic version of the new version.

Destructuring the Response

data CreateApplicationVersionResponse

See: newCreateApplicationVersionResponse smart constructor.

Constructors

CreateApplicationVersionResponse'

Fields

creationTime :: Maybe Text

The date and time this resource was created.

resourcesSupported :: Maybe Bool

Whether all of the AWS resources contained in this application are supported in the region in which it is being retrieved.

requiredCapabilities :: Maybe [Capability]

A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS::TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

parameterDefinitions :: Maybe [ParameterDefinition]

An array of parameter types supported by the application.

semanticVersion :: Maybe Text

The semantic version of the application:

https://semver.org/

sourceCodeUrl :: Maybe Text

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

applicationId :: Maybe Text

The application Amazon Resource Name (ARN).

templateUrl :: Maybe Text

A link to the packaged AWS SAM template of your application.

sourceCodeArchiveUrl :: Maybe Text

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

httpStatus :: Int

The response's http status code.

newCreateApplicationVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateApplicationVersionResponse'

-> CreateApplicationVersionResponse

Create a value of CreateApplicationVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CreateApplicationVersionResponse', createApplicationVersionResponse_creationTime - The date and time this resource was created.

$sel:resourcesSupported:CreateApplicationVersionResponse', createApplicationVersionResponse_resourcesSupported - Whether all of the AWS resources contained in this application are supported in the region in which it is being retrieved.

$sel:requiredCapabilities:CreateApplicationVersionResponse', createApplicationVersionResponse_requiredCapabilities - A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS::TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

$sel:parameterDefinitions:CreateApplicationVersionResponse', createApplicationVersionResponse_parameterDefinitions - An array of parameter types supported by the application.

$sel:semanticVersion:CreateApplicationVersion', createApplicationVersionResponse_semanticVersion - The semantic version of the application:

https://semver.org/

$sel:sourceCodeUrl:CreateApplicationVersion', createApplicationVersionResponse_sourceCodeUrl - A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

$sel:applicationId:CreateApplicationVersion', createApplicationVersionResponse_applicationId - The application Amazon Resource Name (ARN).

$sel:templateUrl:CreateApplicationVersion', createApplicationVersionResponse_templateUrl - A link to the packaged AWS SAM template of your application.

$sel:sourceCodeArchiveUrl:CreateApplicationVersion', createApplicationVersionResponse_sourceCodeArchiveUrl - A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

$sel:httpStatus:CreateApplicationVersionResponse', createApplicationVersionResponse_httpStatus - The response's http status code.

Response Lenses

createApplicationVersionResponse_creationTime :: Lens' CreateApplicationVersionResponse (Maybe Text)

The date and time this resource was created.

createApplicationVersionResponse_resourcesSupported :: Lens' CreateApplicationVersionResponse (Maybe Bool)

Whether all of the AWS resources contained in this application are supported in the region in which it is being retrieved.

createApplicationVersionResponse_requiredCapabilities :: Lens' CreateApplicationVersionResponse (Maybe [Capability])

A list of values that you must specify before you can deploy certain applications. Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.

The only valid values are CAPABILITY_IAM, CAPABILITY_NAMED_IAM, CAPABILITY_RESOURCE_POLICY, and CAPABILITY_AUTO_EXPAND.

The following resources require you to specify CAPABILITY_IAM or CAPABILITY_NAMED_IAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITY_IAM or CAPABILITY_NAMED_IAM. If the application contains IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

The following resources require you to specify CAPABILITY_RESOURCE_POLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS::TopicPolicy.

Applications that contain one or more nested applications require you to specify CAPABILITY_AUTO_EXPAND.

If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

createApplicationVersionResponse_parameterDefinitions :: Lens' CreateApplicationVersionResponse (Maybe [ParameterDefinition])

An array of parameter types supported by the application.

createApplicationVersionResponse_semanticVersion :: Lens' CreateApplicationVersionResponse (Maybe Text)

The semantic version of the application:

https://semver.org/

createApplicationVersionResponse_sourceCodeUrl :: Lens' CreateApplicationVersionResponse (Maybe Text)

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

createApplicationVersionResponse_applicationId :: Lens' CreateApplicationVersionResponse (Maybe Text)

The application Amazon Resource Name (ARN).

createApplicationVersionResponse_templateUrl :: Lens' CreateApplicationVersionResponse (Maybe Text)

A link to the packaged AWS SAM template of your application.

createApplicationVersionResponse_sourceCodeArchiveUrl :: Lens' CreateApplicationVersionResponse (Maybe Text)

A link to the S3 object that contains the ZIP archive of the source code for this version of your application.

Maximum size 50 MB

createApplicationVersionResponse_httpStatus :: Lens' CreateApplicationVersionResponse Int

The response's http status code.