Amazonka.ServerlessApplicationRepository.Types.VersionSummary

Description

Documentation

data VersionSummary

An application version summary.

See: newVersionSummary smart constructor.

Constructors

VersionSummary'

Fields

sourceCodeUrl :: Maybe Text

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

creationTime :: Text

The date and time this resource was created.

applicationId :: Text

The application Amazon Resource Name (ARN).

semanticVersion :: Text

The semantic version of the application:

https://semver.org/

newVersionSummary

Arguments

:: Text
$sel:creationTime:VersionSummary'

-> Text
$sel:applicationId:VersionSummary'

-> Text
$sel:semanticVersion:VersionSummary'

-> VersionSummary

Create a value of VersionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceCodeUrl:VersionSummary', versionSummary_sourceCodeUrl - A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

$sel:creationTime:VersionSummary', versionSummary_creationTime - The date and time this resource was created.

$sel:applicationId:VersionSummary', versionSummary_applicationId - The application Amazon Resource Name (ARN).

$sel:semanticVersion:VersionSummary', versionSummary_semanticVersion - The semantic version of the application:

https://semver.org/

versionSummary_sourceCodeUrl :: Lens' VersionSummary (Maybe Text)

A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

versionSummary_creationTime :: Lens' VersionSummary Text

The date and time this resource was created.

versionSummary_applicationId :: Lens' VersionSummary Text

The application Amazon Resource Name (ARN).

versionSummary_semanticVersion :: Lens' VersionSummary Text

The semantic version of the application:

https://semver.org/