Amazonka.ServiceCatalog.BatchDisassociateServiceActionFromProvisioningArtifact

Description

Disassociates a batch of self-service actions from the specified provisioning artifact.

Creating a Request

data BatchDisassociateServiceActionFromProvisioningArtifact

See: newBatchDisassociateServiceActionFromProvisioningArtifact smart constructor.

Constructors

BatchDisassociateServiceActionFromProvisioningArtifact'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

serviceActionAssociations :: NonEmpty ServiceActionAssociation

One or more associations, each consisting of the Action ID, the Product ID, and the Provisioning Artifact ID.

newBatchDisassociateServiceActionFromProvisioningArtifact

Arguments

:: NonEmpty ServiceActionAssociation
$sel:serviceActionAssociations:BatchDisassociateServiceActionFromProvisioningArtifact'

-> BatchDisassociateServiceActionFromProvisioningArtifact

Create a value of BatchDisassociateServiceActionFromProvisioningArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:BatchDisassociateServiceActionFromProvisioningArtifact', batchDisassociateServiceActionFromProvisioningArtifact_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:serviceActionAssociations:BatchDisassociateServiceActionFromProvisioningArtifact', batchDisassociateServiceActionFromProvisioningArtifact_serviceActionAssociations - One or more associations, each consisting of the Action ID, the Product ID, and the Provisioning Artifact ID.

Request Lenses

batchDisassociateServiceActionFromProvisioningArtifact_acceptLanguage :: Lens' BatchDisassociateServiceActionFromProvisioningArtifact (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

batchDisassociateServiceActionFromProvisioningArtifact_serviceActionAssociations :: Lens' BatchDisassociateServiceActionFromProvisioningArtifact (NonEmpty ServiceActionAssociation)

One or more associations, each consisting of the Action ID, the Product ID, and the Provisioning Artifact ID.

Destructuring the Response

data BatchDisassociateServiceActionFromProvisioningArtifactResponse

See: newBatchDisassociateServiceActionFromProvisioningArtifactResponse smart constructor.

Constructors

BatchDisassociateServiceActionFromProvisioningArtifactResponse'

Fields

failedServiceActionAssociations :: Maybe [FailedServiceActionAssociation]

An object that contains a list of errors, along with information to help you identify the self-service action.

httpStatus :: Int

The response's http status code.

newBatchDisassociateServiceActionFromProvisioningArtifactResponse

Arguments

:: Int
$sel:httpStatus:BatchDisassociateServiceActionFromProvisioningArtifactResponse'

-> BatchDisassociateServiceActionFromProvisioningArtifactResponse

Create a value of BatchDisassociateServiceActionFromProvisioningArtifactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedServiceActionAssociations:BatchDisassociateServiceActionFromProvisioningArtifactResponse', batchDisassociateServiceActionFromProvisioningArtifactResponse_failedServiceActionAssociations - An object that contains a list of errors, along with information to help you identify the self-service action.

$sel:httpStatus:BatchDisassociateServiceActionFromProvisioningArtifactResponse', batchDisassociateServiceActionFromProvisioningArtifactResponse_httpStatus - The response's http status code.

Response Lenses

batchDisassociateServiceActionFromProvisioningArtifactResponse_failedServiceActionAssociations :: Lens' BatchDisassociateServiceActionFromProvisioningArtifactResponse (Maybe [FailedServiceActionAssociation])

An object that contains a list of errors, along with information to help you identify the self-service action.

batchDisassociateServiceActionFromProvisioningArtifactResponse_httpStatus :: Lens' BatchDisassociateServiceActionFromProvisioningArtifactResponse Int

The response's http status code.