Amazonka.ServiceCatalog.CreateConstraint

Description

Creates a constraint.

A delegated admin is authorized to invoke this command.

Creating a Request

data CreateConstraint

See: newCreateConstraint smart constructor.

Constructors

CreateConstraint'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

description :: Maybe Text

The description of the constraint.

portfolioId :: Text

The portfolio identifier.

productId :: Text

The product identifier.

parameters :: Text

The constraint parameters, in JSON format. The syntax depends on the constraint type as follows:

LAUNCH

You are required to specify either the RoleArn or the LocalRoleName but can't use both.
Specify the RoleArn property as follows:

{"RoleArn" : "arn:aws:iam::123456789012:role/LaunchRole"}

Specify the LocalRoleName property as follows:

{"LocalRoleName": "SCBasicLaunchRole"}

If you specify the LocalRoleName property, when an account uses the launch constraint, the IAM role with that name in the account will be used. This allows launch-role constraints to be account-agnostic so the administrator can create fewer resources per shared account.

The given role name must exist in the account used to create the launch constraint and the account of the user who launches a product with this launch constraint.

You cannot have both a LAUNCH and a STACKSET constraint.

You also cannot have more than one LAUNCH constraint on a product and portfolio.

NOTIFICATION

Specify the NotificationArns property as follows:
{"NotificationArns" : ["arn:aws:sns:us-east-1:123456789012:Topic"]}

RESOURCE_UPDATE

Specify the TagUpdatesOnProvisionedProduct property as follows:
{"Version":"2.0","Properties":{"TagUpdateOnProvisionedProduct":"String"}}

The TagUpdatesOnProvisionedProduct property accepts a string value of ALLOWED or NOT_ALLOWED.

STACKSET

Specify the Parameters property as follows:
{"Version": "String", "Properties": {"AccountList": [ "String" ], "RegionList": [ "String" ], "AdminRole": "String", "ExecutionRole": "String"}}

You cannot have both a LAUNCH and a STACKSET constraint.

You also cannot have more than one STACKSET constraint on a product and portfolio.

Products with a STACKSET constraint will launch an AWS CloudFormation stack set.

TEMPLATE

Specify the Rules property. For more information, see Template Constraint Rules.

type' :: Text

The type of constraint.

LAUNCH

NOTIFICATION

RESOURCE_UPDATE

STACKSET

TEMPLATE

idempotencyToken :: Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

newCreateConstraint

Arguments

:: Text
$sel:portfolioId:CreateConstraint'

-> Text
$sel:productId:CreateConstraint'

-> Text
$sel:parameters:CreateConstraint'

-> Text
$sel:type':CreateConstraint'

-> Text
$sel:idempotencyToken:CreateConstraint'

-> CreateConstraint

Create a value of CreateConstraint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:CreateConstraint', createConstraint_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:description:CreateConstraint', createConstraint_description - The description of the constraint.

$sel:portfolioId:CreateConstraint', createConstraint_portfolioId - The portfolio identifier.

$sel:productId:CreateConstraint', createConstraint_productId - The product identifier.

$sel:parameters:CreateConstraint', createConstraint_parameters - The constraint parameters, in JSON format. The syntax depends on the constraint type as follows:

LAUNCH

You are required to specify either the RoleArn or the LocalRoleName but can't use both.
Specify the RoleArn property as follows:

{"RoleArn" : "arn:aws:iam::123456789012:role/LaunchRole"}

Specify the LocalRoleName property as follows:

{"LocalRoleName": "SCBasicLaunchRole"}

If you specify the LocalRoleName property, when an account uses the launch constraint, the IAM role with that name in the account will be used. This allows launch-role constraints to be account-agnostic so the administrator can create fewer resources per shared account.

The given role name must exist in the account used to create the launch constraint and the account of the user who launches a product with this launch constraint.

You cannot have both a LAUNCH and a STACKSET constraint.

You also cannot have more than one LAUNCH constraint on a product and portfolio.

NOTIFICATION

Specify the NotificationArns property as follows:
{"NotificationArns" : ["arn:aws:sns:us-east-1:123456789012:Topic"]}

RESOURCE_UPDATE

Specify the TagUpdatesOnProvisionedProduct property as follows:
{"Version":"2.0","Properties":{"TagUpdateOnProvisionedProduct":"String"}}

The TagUpdatesOnProvisionedProduct property accepts a string value of ALLOWED or NOT_ALLOWED.

STACKSET

Specify the Parameters property as follows:
{"Version": "String", "Properties": {"AccountList": [ "String" ], "RegionList": [ "String" ], "AdminRole": "String", "ExecutionRole": "String"}}

You cannot have both a LAUNCH and a STACKSET constraint.

You also cannot have more than one STACKSET constraint on a product and portfolio.

Products with a STACKSET constraint will launch an AWS CloudFormation stack set.

TEMPLATE

Specify the Rules property. For more information, see Template Constraint Rules.

$sel:type':CreateConstraint', createConstraint_type - The type of constraint.

LAUNCH

NOTIFICATION

RESOURCE_UPDATE

STACKSET

TEMPLATE

$sel:idempotencyToken:CreateConstraint', createConstraint_idempotencyToken - A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

Request Lenses

createConstraint_acceptLanguage :: Lens' CreateConstraint (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

createConstraint_description :: Lens' CreateConstraint (Maybe Text)

The description of the constraint.

createConstraint_portfolioId :: Lens' CreateConstraint Text

The portfolio identifier.

createConstraint_productId :: Lens' CreateConstraint Text

The product identifier.

createConstraint_parameters :: Lens' CreateConstraint Text

The constraint parameters, in JSON format. The syntax depends on the constraint type as follows:

LAUNCH

You are required to specify either the RoleArn or the LocalRoleName but can't use both.
Specify the RoleArn property as follows:

{"RoleArn" : "arn:aws:iam::123456789012:role/LaunchRole"}

Specify the LocalRoleName property as follows:

{"LocalRoleName": "SCBasicLaunchRole"}

If you specify the LocalRoleName property, when an account uses the launch constraint, the IAM role with that name in the account will be used. This allows launch-role constraints to be account-agnostic so the administrator can create fewer resources per shared account.

The given role name must exist in the account used to create the launch constraint and the account of the user who launches a product with this launch constraint.

You cannot have both a LAUNCH and a STACKSET constraint.

You also cannot have more than one LAUNCH constraint on a product and portfolio.

NOTIFICATION

Specify the NotificationArns property as follows:
{"NotificationArns" : ["arn:aws:sns:us-east-1:123456789012:Topic"]}

RESOURCE_UPDATE

Specify the TagUpdatesOnProvisionedProduct property as follows:
{"Version":"2.0","Properties":{"TagUpdateOnProvisionedProduct":"String"}}

The TagUpdatesOnProvisionedProduct property accepts a string value of ALLOWED or NOT_ALLOWED.

STACKSET

Specify the Parameters property as follows:
{"Version": "String", "Properties": {"AccountList": [ "String" ], "RegionList": [ "String" ], "AdminRole": "String", "ExecutionRole": "String"}}

You cannot have both a LAUNCH and a STACKSET constraint.

You also cannot have more than one STACKSET constraint on a product and portfolio.

Products with a STACKSET constraint will launch an AWS CloudFormation stack set.

TEMPLATE

Specify the Rules property. For more information, see Template Constraint Rules.

createConstraint_type :: Lens' CreateConstraint Text

The type of constraint.

LAUNCH

NOTIFICATION

RESOURCE_UPDATE

STACKSET

TEMPLATE

createConstraint_idempotencyToken :: Lens' CreateConstraint Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

Destructuring the Response

data CreateConstraintResponse

See: newCreateConstraintResponse smart constructor.

Constructors

CreateConstraintResponse'

Fields

status :: Maybe RequestStatus

The status of the current request.

constraintDetail :: Maybe ConstraintDetail

Information about the constraint.

constraintParameters :: Maybe Text

The constraint parameters.

httpStatus :: Int

The response's http status code.

newCreateConstraintResponse

Arguments

:: Int
$sel:httpStatus:CreateConstraintResponse'

-> CreateConstraintResponse

Create a value of CreateConstraintResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateConstraintResponse', createConstraintResponse_status - The status of the current request.

$sel:constraintDetail:CreateConstraintResponse', createConstraintResponse_constraintDetail - Information about the constraint.

$sel:constraintParameters:CreateConstraintResponse', createConstraintResponse_constraintParameters - The constraint parameters.

$sel:httpStatus:CreateConstraintResponse', createConstraintResponse_httpStatus - The response's http status code.

Response Lenses

createConstraintResponse_status :: Lens' CreateConstraintResponse (Maybe RequestStatus)

The status of the current request.

createConstraintResponse_constraintDetail :: Lens' CreateConstraintResponse (Maybe ConstraintDetail)

Information about the constraint.

createConstraintResponse_constraintParameters :: Lens' CreateConstraintResponse (Maybe Text)

The constraint parameters.

createConstraintResponse_httpStatus :: Lens' CreateConstraintResponse Int

The response's http status code.