Amazonka.ServiceCatalog.CreatePortfolioShare

Description

Shares the specified portfolio with the specified account or organization node. Shares to an organization node can only be created by the management account of an organization or by a delegated administrator. You can share portfolios to an organization, an organizational unit, or a specific account.

Note that if a delegated admin is de-registered, they can no longer create portfolio shares.

AWSOrganizationsAccess must be enabled in order to create a portfolio share to an organization node.

You can't share a shared resource, including portfolios that contain a shared product.

If the portfolio share with the specified account or organization node already exists, this action will have no effect and will not return an error. To update an existing share, you must use the UpdatePortfolioShare API instead.

Creating a Request

data CreatePortfolioShare

See: newCreatePortfolioShare smart constructor.

Constructors

CreatePortfolioShare'

Fields

accountId :: Maybe Text

The AWS account ID. For example, 123456789012.

shareTagOptions :: Maybe Bool

Enables or disables TagOptions sharing when creating the portfolio share. If this flag is not provided, TagOptions sharing is disabled.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

organizationNode :: Maybe OrganizationNode

The organization node to whom you are going to share. If OrganizationNode is passed in, PortfolioShare will be created for the node an ListOrganizationPortfolioAccessd its children (when applies), and a PortfolioShareToken will be returned in the output in order for the administrator to monitor the status of the PortfolioShare creation process.

portfolioId :: Text

The portfolio identifier.

newCreatePortfolioShare

Arguments

:: Text
$sel:portfolioId:CreatePortfolioShare'

-> CreatePortfolioShare

Create a value of CreatePortfolioShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:CreatePortfolioShare', createPortfolioShare_accountId - The AWS account ID. For example, 123456789012.

$sel:shareTagOptions:CreatePortfolioShare', createPortfolioShare_shareTagOptions - Enables or disables TagOptions sharing when creating the portfolio share. If this flag is not provided, TagOptions sharing is disabled.

$sel:acceptLanguage:CreatePortfolioShare', createPortfolioShare_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:organizationNode:CreatePortfolioShare', createPortfolioShare_organizationNode - The organization node to whom you are going to share. If OrganizationNode is passed in, PortfolioShare will be created for the node an ListOrganizationPortfolioAccessd its children (when applies), and a PortfolioShareToken will be returned in the output in order for the administrator to monitor the status of the PortfolioShare creation process.

$sel:portfolioId:CreatePortfolioShare', createPortfolioShare_portfolioId - The portfolio identifier.

Request Lenses

createPortfolioShare_accountId :: Lens' CreatePortfolioShare (Maybe Text)

The AWS account ID. For example, 123456789012.

createPortfolioShare_shareTagOptions :: Lens' CreatePortfolioShare (Maybe Bool)

Enables or disables TagOptions sharing when creating the portfolio share. If this flag is not provided, TagOptions sharing is disabled.

createPortfolioShare_acceptLanguage :: Lens' CreatePortfolioShare (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

createPortfolioShare_organizationNode :: Lens' CreatePortfolioShare (Maybe OrganizationNode)

The organization node to whom you are going to share. If OrganizationNode is passed in, PortfolioShare will be created for the node an ListOrganizationPortfolioAccessd its children (when applies), and a PortfolioShareToken will be returned in the output in order for the administrator to monitor the status of the PortfolioShare creation process.

createPortfolioShare_portfolioId :: Lens' CreatePortfolioShare Text

The portfolio identifier.

Destructuring the Response

data CreatePortfolioShareResponse

See: newCreatePortfolioShareResponse smart constructor.

Constructors

CreatePortfolioShareResponse'

Fields

portfolioShareToken :: Maybe Text

The portfolio shares a unique identifier that only returns if the portfolio is shared to an organization node.

httpStatus :: Int

The response's http status code.

newCreatePortfolioShareResponse

Arguments

:: Int
$sel:httpStatus:CreatePortfolioShareResponse'

-> CreatePortfolioShareResponse

Create a value of CreatePortfolioShareResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portfolioShareToken:CreatePortfolioShareResponse', createPortfolioShareResponse_portfolioShareToken - The portfolio shares a unique identifier that only returns if the portfolio is shared to an organization node.

$sel:httpStatus:CreatePortfolioShareResponse', createPortfolioShareResponse_httpStatus - The response's http status code.

Response Lenses

createPortfolioShareResponse_portfolioShareToken :: Lens' CreatePortfolioShareResponse (Maybe Text)

The portfolio shares a unique identifier that only returns if the portfolio is shared to an organization node.

createPortfolioShareResponse_httpStatus :: Lens' CreatePortfolioShareResponse Int

The response's http status code.