Amazonka.ServiceCatalog.CreateProvisioningArtifact

Description

Creates a provisioning artifact (also known as a version) for the specified product.

You cannot create a provisioning artifact for a product that was shared with you.

The user or role that performs this operation must have the cloudformation:GetTemplate IAM policy permission. This policy permission is required when using the ImportFromPhysicalId template source in the information data section.

Creating a Request

data CreateProvisioningArtifact

See: newCreateProvisioningArtifact smart constructor.

Constructors

CreateProvisioningArtifact'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

productId :: Text

The product identifier.

parameters :: ProvisioningArtifactProperties

The configuration for the provisioning artifact.

idempotencyToken :: Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

newCreateProvisioningArtifact

Arguments

:: Text
$sel:productId:CreateProvisioningArtifact'

-> ProvisioningArtifactProperties
$sel:parameters:CreateProvisioningArtifact'

-> Text
$sel:idempotencyToken:CreateProvisioningArtifact'

-> CreateProvisioningArtifact

Create a value of CreateProvisioningArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:CreateProvisioningArtifact', createProvisioningArtifact_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:productId:CreateProvisioningArtifact', createProvisioningArtifact_productId - The product identifier.

$sel:parameters:CreateProvisioningArtifact', createProvisioningArtifact_parameters - The configuration for the provisioning artifact.

$sel:idempotencyToken:CreateProvisioningArtifact', createProvisioningArtifact_idempotencyToken - A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

Request Lenses

createProvisioningArtifact_acceptLanguage :: Lens' CreateProvisioningArtifact (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

createProvisioningArtifact_productId :: Lens' CreateProvisioningArtifact Text

The product identifier.

createProvisioningArtifact_parameters :: Lens' CreateProvisioningArtifact ProvisioningArtifactProperties

The configuration for the provisioning artifact.

createProvisioningArtifact_idempotencyToken :: Lens' CreateProvisioningArtifact Text

A unique identifier that you provide to ensure idempotency. If multiple requests differ only by the idempotency token, the same response is returned for each repeated request.

Destructuring the Response

data CreateProvisioningArtifactResponse

See: newCreateProvisioningArtifactResponse smart constructor.

Constructors

CreateProvisioningArtifactResponse'

Fields

status :: Maybe RequestStatus

The status of the current request.

info :: Maybe (HashMap Text Text)

Specify the template source with one of the following options, but not both. Keys accepted: [ LoadTemplateFromURL, ImportFromPhysicalId ].

The URL of the CloudFormation template in Amazon S3, in JSON format.

LoadTemplateFromURL

Use the URL of the CloudFormation template in Amazon S3 in JSON format.

ImportFromPhysicalId

Use the physical id of the resource that contains the template; currently supports CloudFormation stack ARN.

provisioningArtifactDetail :: Maybe ProvisioningArtifactDetail

Information about the provisioning artifact.

httpStatus :: Int

The response's http status code.

newCreateProvisioningArtifactResponse

Arguments

:: Int
$sel:httpStatus:CreateProvisioningArtifactResponse'

-> CreateProvisioningArtifactResponse

Create a value of CreateProvisioningArtifactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateProvisioningArtifactResponse', createProvisioningArtifactResponse_status - The status of the current request.

$sel:info:CreateProvisioningArtifactResponse', createProvisioningArtifactResponse_info - Specify the template source with one of the following options, but not both. Keys accepted: [ LoadTemplateFromURL, ImportFromPhysicalId ].

The URL of the CloudFormation template in Amazon S3, in JSON format.

LoadTemplateFromURL

Use the URL of the CloudFormation template in Amazon S3 in JSON format.

ImportFromPhysicalId

Use the physical id of the resource that contains the template; currently supports CloudFormation stack ARN.

$sel:provisioningArtifactDetail:CreateProvisioningArtifactResponse', createProvisioningArtifactResponse_provisioningArtifactDetail - Information about the provisioning artifact.

$sel:httpStatus:CreateProvisioningArtifactResponse', createProvisioningArtifactResponse_httpStatus - The response's http status code.

Response Lenses

createProvisioningArtifactResponse_status :: Lens' CreateProvisioningArtifactResponse (Maybe RequestStatus)

The status of the current request.

createProvisioningArtifactResponse_info :: Lens' CreateProvisioningArtifactResponse (Maybe (HashMap Text Text))

Specify the template source with one of the following options, but not both. Keys accepted: [ LoadTemplateFromURL, ImportFromPhysicalId ].

The URL of the CloudFormation template in Amazon S3, in JSON format.

LoadTemplateFromURL

Use the URL of the CloudFormation template in Amazon S3 in JSON format.

ImportFromPhysicalId

Use the physical id of the resource that contains the template; currently supports CloudFormation stack ARN.

createProvisioningArtifactResponse_provisioningArtifactDetail :: Lens' CreateProvisioningArtifactResponse (Maybe ProvisioningArtifactDetail)

Information about the provisioning artifact.

createProvisioningArtifactResponse_httpStatus :: Lens' CreateProvisioningArtifactResponse Int

The response's http status code.