Amazonka.ServiceCatalog.DeletePortfolioShare

Description

Stops sharing the specified portfolio with the specified account or organization node. Shares to an organization node can only be deleted by the management account of an organization or by a delegated administrator.

Note that if a delegated admin is de-registered, portfolio shares created from that account are removed.

Creating a Request

data DeletePortfolioShare

See: newDeletePortfolioShare smart constructor.

Constructors

DeletePortfolioShare'

Fields

accountId :: Maybe Text

The AWS account ID.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

organizationNode :: Maybe OrganizationNode

The organization node to whom you are going to stop sharing.

portfolioId :: Text

The portfolio identifier.

newDeletePortfolioShare

Arguments

:: Text
$sel:portfolioId:DeletePortfolioShare'

-> DeletePortfolioShare

Create a value of DeletePortfolioShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:DeletePortfolioShare', deletePortfolioShare_accountId - The AWS account ID.

$sel:acceptLanguage:DeletePortfolioShare', deletePortfolioShare_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:organizationNode:DeletePortfolioShare', deletePortfolioShare_organizationNode - The organization node to whom you are going to stop sharing.

$sel:portfolioId:DeletePortfolioShare', deletePortfolioShare_portfolioId - The portfolio identifier.

Request Lenses

deletePortfolioShare_accountId :: Lens' DeletePortfolioShare (Maybe Text)

The AWS account ID.

deletePortfolioShare_acceptLanguage :: Lens' DeletePortfolioShare (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

deletePortfolioShare_organizationNode :: Lens' DeletePortfolioShare (Maybe OrganizationNode)

The organization node to whom you are going to stop sharing.

deletePortfolioShare_portfolioId :: Lens' DeletePortfolioShare Text

The portfolio identifier.

Destructuring the Response

data DeletePortfolioShareResponse

See: newDeletePortfolioShareResponse smart constructor.

Constructors

DeletePortfolioShareResponse'

Fields

portfolioShareToken :: Maybe Text

The portfolio share unique identifier. This will only be returned if delete is made to an organization node.

httpStatus :: Int

The response's http status code.

newDeletePortfolioShareResponse

Arguments

:: Int
$sel:httpStatus:DeletePortfolioShareResponse'

-> DeletePortfolioShareResponse

Create a value of DeletePortfolioShareResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portfolioShareToken:DeletePortfolioShareResponse', deletePortfolioShareResponse_portfolioShareToken - The portfolio share unique identifier. This will only be returned if delete is made to an organization node.

$sel:httpStatus:DeletePortfolioShareResponse', deletePortfolioShareResponse_httpStatus - The response's http status code.

Response Lenses

deletePortfolioShareResponse_portfolioShareToken :: Lens' DeletePortfolioShareResponse (Maybe Text)

The portfolio share unique identifier. This will only be returned if delete is made to an organization node.

deletePortfolioShareResponse_httpStatus :: Lens' DeletePortfolioShareResponse Int

The response's http status code.