Amazonka.ServiceCatalog.DescribeConstraint

Description

Gets information about the specified constraint.

Creating a Request

data DescribeConstraint

See: newDescribeConstraint smart constructor.

Constructors

DescribeConstraint'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

id :: Text

The identifier of the constraint.

newDescribeConstraint

Arguments

:: Text
$sel:id:DescribeConstraint'

-> DescribeConstraint

Create a value of DescribeConstraint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:DescribeConstraint', describeConstraint_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:id:DescribeConstraint', describeConstraint_id - The identifier of the constraint.

Request Lenses

describeConstraint_acceptLanguage :: Lens' DescribeConstraint (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

describeConstraint_id :: Lens' DescribeConstraint Text

The identifier of the constraint.

Destructuring the Response

data DescribeConstraintResponse

See: newDescribeConstraintResponse smart constructor.

Constructors

DescribeConstraintResponse'

Fields

status :: Maybe RequestStatus

The status of the current request.

constraintDetail :: Maybe ConstraintDetail

Information about the constraint.

constraintParameters :: Maybe Text

The constraint parameters.

httpStatus :: Int

The response's http status code.

newDescribeConstraintResponse

Arguments

:: Int
$sel:httpStatus:DescribeConstraintResponse'

-> DescribeConstraintResponse

Create a value of DescribeConstraintResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeConstraintResponse', describeConstraintResponse_status - The status of the current request.

$sel:constraintDetail:DescribeConstraintResponse', describeConstraintResponse_constraintDetail - Information about the constraint.

$sel:constraintParameters:DescribeConstraintResponse', describeConstraintResponse_constraintParameters - The constraint parameters.

$sel:httpStatus:DescribeConstraintResponse', describeConstraintResponse_httpStatus - The response's http status code.

Response Lenses

describeConstraintResponse_status :: Lens' DescribeConstraintResponse (Maybe RequestStatus)

The status of the current request.

describeConstraintResponse_constraintDetail :: Lens' DescribeConstraintResponse (Maybe ConstraintDetail)

Information about the constraint.

describeConstraintResponse_constraintParameters :: Lens' DescribeConstraintResponse (Maybe Text)

The constraint parameters.

describeConstraintResponse_httpStatus :: Lens' DescribeConstraintResponse Int

The response's http status code.