Amazonka.ServiceCatalog.DescribePortfolio

Description

Gets information about the specified portfolio.

A delegated admin is authorized to invoke this command.

Creating a Request

data DescribePortfolio

See: newDescribePortfolio smart constructor.

Constructors

DescribePortfolio'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

id :: Text

The portfolio identifier.

newDescribePortfolio

Arguments

:: Text
$sel:id:DescribePortfolio'

-> DescribePortfolio

Create a value of DescribePortfolio with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:DescribePortfolio', describePortfolio_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:id:DescribePortfolio', describePortfolio_id - The portfolio identifier.

Request Lenses

describePortfolio_acceptLanguage :: Lens' DescribePortfolio (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

describePortfolio_id :: Lens' DescribePortfolio Text

The portfolio identifier.

Destructuring the Response

data DescribePortfolioResponse

See: newDescribePortfolioResponse smart constructor.

Constructors

DescribePortfolioResponse'

Fields

portfolioDetail :: Maybe PortfolioDetail

Information about the portfolio.

tagOptions :: Maybe [TagOptionDetail]

Information about the TagOptions associated with the portfolio.

budgets :: Maybe [BudgetDetail]

Information about the associated budgets.

tags :: Maybe [Tag]

Information about the tags associated with the portfolio.

httpStatus :: Int

The response's http status code.

newDescribePortfolioResponse

Arguments

:: Int
$sel:httpStatus:DescribePortfolioResponse'

-> DescribePortfolioResponse

Create a value of DescribePortfolioResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portfolioDetail:DescribePortfolioResponse', describePortfolioResponse_portfolioDetail - Information about the portfolio.

$sel:tagOptions:DescribePortfolioResponse', describePortfolioResponse_tagOptions - Information about the TagOptions associated with the portfolio.

$sel:budgets:DescribePortfolioResponse', describePortfolioResponse_budgets - Information about the associated budgets.

$sel:tags:DescribePortfolioResponse', describePortfolioResponse_tags - Information about the tags associated with the portfolio.

$sel:httpStatus:DescribePortfolioResponse', describePortfolioResponse_httpStatus - The response's http status code.

Response Lenses

describePortfolioResponse_portfolioDetail :: Lens' DescribePortfolioResponse (Maybe PortfolioDetail)

Information about the portfolio.

describePortfolioResponse_tagOptions :: Lens' DescribePortfolioResponse (Maybe [TagOptionDetail])

Information about the TagOptions associated with the portfolio.

describePortfolioResponse_budgets :: Lens' DescribePortfolioResponse (Maybe [BudgetDetail])

Information about the associated budgets.

describePortfolioResponse_tags :: Lens' DescribePortfolioResponse (Maybe [Tag])

Information about the tags associated with the portfolio.

describePortfolioResponse_httpStatus :: Lens' DescribePortfolioResponse Int

The response's http status code.