Amazonka.ServiceCatalog.DescribeProvisionedProduct

Description

Gets information about the specified provisioned product.

Creating a Request

data DescribeProvisionedProduct

DescribeProvisionedProductAPI input structure. AcceptLanguage - [Optional] The language code for localization. Id - [Optional] The provisioned product identifier. Name - [Optional] Another provisioned product identifier. Customers must provide either Id or Name.

See: newDescribeProvisionedProduct smart constructor.

Constructors

DescribeProvisionedProduct'

Fields

name :: Maybe Text

The name of the provisioned product. You must provide the name or ID, but not both.

If you do not provide a name or ID, or you provide both name and ID, an InvalidParametersException will occur.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

id :: Maybe Text

The provisioned product identifier. You must provide the name or ID, but not both.

If you do not provide a name or ID, or you provide both name and ID, an InvalidParametersException will occur.

newDescribeProvisionedProduct :: DescribeProvisionedProduct

Create a value of DescribeProvisionedProduct with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeProvisionedProduct', describeProvisionedProduct_name - The name of the provisioned product. You must provide the name or ID, but not both.

If you do not provide a name or ID, or you provide both name and ID, an InvalidParametersException will occur.

$sel:acceptLanguage:DescribeProvisionedProduct', describeProvisionedProduct_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:id:DescribeProvisionedProduct', describeProvisionedProduct_id - The provisioned product identifier. You must provide the name or ID, but not both.

If you do not provide a name or ID, or you provide both name and ID, an InvalidParametersException will occur.

Request Lenses

describeProvisionedProduct_name :: Lens' DescribeProvisionedProduct (Maybe Text)

The name of the provisioned product. You must provide the name or ID, but not both.

If you do not provide a name or ID, or you provide both name and ID, an InvalidParametersException will occur.

describeProvisionedProduct_acceptLanguage :: Lens' DescribeProvisionedProduct (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

describeProvisionedProduct_id :: Lens' DescribeProvisionedProduct (Maybe Text)

The provisioned product identifier. You must provide the name or ID, but not both.

If you do not provide a name or ID, or you provide both name and ID, an InvalidParametersException will occur.

Destructuring the Response

data DescribeProvisionedProductResponse

See: newDescribeProvisionedProductResponse smart constructor.

Constructors

DescribeProvisionedProductResponse'

Fields

provisionedProductDetail :: Maybe ProvisionedProductDetail

Information about the provisioned product.

cloudWatchDashboards :: Maybe [CloudWatchDashboard]

Any CloudWatch dashboards that were created when provisioning the product.

httpStatus :: Int

The response's http status code.

newDescribeProvisionedProductResponse

Arguments

:: Int
$sel:httpStatus:DescribeProvisionedProductResponse'

-> DescribeProvisionedProductResponse

Create a value of DescribeProvisionedProductResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisionedProductDetail:DescribeProvisionedProductResponse', describeProvisionedProductResponse_provisionedProductDetail - Information about the provisioned product.

$sel:cloudWatchDashboards:DescribeProvisionedProductResponse', describeProvisionedProductResponse_cloudWatchDashboards - Any CloudWatch dashboards that were created when provisioning the product.

$sel:httpStatus:DescribeProvisionedProductResponse', describeProvisionedProductResponse_httpStatus - The response's http status code.

Response Lenses

describeProvisionedProductResponse_provisionedProductDetail :: Lens' DescribeProvisionedProductResponse (Maybe ProvisionedProductDetail)

Information about the provisioned product.

describeProvisionedProductResponse_cloudWatchDashboards :: Lens' DescribeProvisionedProductResponse (Maybe [CloudWatchDashboard])

Any CloudWatch dashboards that were created when provisioning the product.

describeProvisionedProductResponse_httpStatus :: Lens' DescribeProvisionedProductResponse Int

The response's http status code.