Amazonka.ServiceCatalog.ListOrganizationPortfolioAccess

Description

Lists the organization nodes that have access to the specified portfolio. This API can only be called by the management account in the organization or by a delegated admin.

If a delegated admin is de-registered, they can no longer perform this operation.

This operation returns paginated results.

Creating a Request

data ListOrganizationPortfolioAccess

See: newListOrganizationPortfolioAccess smart constructor.

Constructors

ListOrganizationPortfolioAccess'

Fields

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

pageSize :: Maybe Natural

The maximum number of items to return with this call.

portfolioId :: Text

The portfolio identifier. For example, port-2abcdext3y5fk.

organizationNodeType :: OrganizationNodeType

The organization node type that will be returned in the output.

ORGANIZATION - Organization that has access to the portfolio.

ORGANIZATIONAL_UNIT - Organizational unit that has access to the portfolio within your organization.

ACCOUNT - Account that has access to the portfolio within your organization.

newListOrganizationPortfolioAccess

Arguments

:: Text
$sel:portfolioId:ListOrganizationPortfolioAccess'

-> OrganizationNodeType
$sel:organizationNodeType:ListOrganizationPortfolioAccess'

-> ListOrganizationPortfolioAccess

Create a value of ListOrganizationPortfolioAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceptLanguage:ListOrganizationPortfolioAccess', listOrganizationPortfolioAccess_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:pageToken:ListOrganizationPortfolioAccess', listOrganizationPortfolioAccess_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:pageSize:ListOrganizationPortfolioAccess', listOrganizationPortfolioAccess_pageSize - The maximum number of items to return with this call.

$sel:portfolioId:ListOrganizationPortfolioAccess', listOrganizationPortfolioAccess_portfolioId - The portfolio identifier. For example, port-2abcdext3y5fk.

$sel:organizationNodeType:ListOrganizationPortfolioAccess', listOrganizationPortfolioAccess_organizationNodeType - The organization node type that will be returned in the output.

ORGANIZATION - Organization that has access to the portfolio.

ORGANIZATIONAL_UNIT - Organizational unit that has access to the portfolio within your organization.

ACCOUNT - Account that has access to the portfolio within your organization.

Request Lenses

listOrganizationPortfolioAccess_acceptLanguage :: Lens' ListOrganizationPortfolioAccess (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

listOrganizationPortfolioAccess_pageToken :: Lens' ListOrganizationPortfolioAccess (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

listOrganizationPortfolioAccess_pageSize :: Lens' ListOrganizationPortfolioAccess (Maybe Natural)

The maximum number of items to return with this call.

listOrganizationPortfolioAccess_portfolioId :: Lens' ListOrganizationPortfolioAccess Text

The portfolio identifier. For example, port-2abcdext3y5fk.

listOrganizationPortfolioAccess_organizationNodeType :: Lens' ListOrganizationPortfolioAccess OrganizationNodeType

The organization node type that will be returned in the output.

ORGANIZATION - Organization that has access to the portfolio.

ORGANIZATIONAL_UNIT - Organizational unit that has access to the portfolio within your organization.

ACCOUNT - Account that has access to the portfolio within your organization.

Destructuring the Response

data ListOrganizationPortfolioAccessResponse

See: newListOrganizationPortfolioAccessResponse smart constructor.

Constructors

ListOrganizationPortfolioAccessResponse'

Fields

nextPageToken :: Maybe Text

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

organizationNodes :: Maybe [OrganizationNode]

Displays information about the organization nodes.

httpStatus :: Int

The response's http status code.

newListOrganizationPortfolioAccessResponse

Arguments

:: Int
$sel:httpStatus:ListOrganizationPortfolioAccessResponse'

-> ListOrganizationPortfolioAccessResponse

Create a value of ListOrganizationPortfolioAccessResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:ListOrganizationPortfolioAccessResponse', listOrganizationPortfolioAccessResponse_nextPageToken - The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

$sel:organizationNodes:ListOrganizationPortfolioAccessResponse', listOrganizationPortfolioAccessResponse_organizationNodes - Displays information about the organization nodes.

$sel:httpStatus:ListOrganizationPortfolioAccessResponse', listOrganizationPortfolioAccessResponse_httpStatus - The response's http status code.

Response Lenses

listOrganizationPortfolioAccessResponse_nextPageToken :: Lens' ListOrganizationPortfolioAccessResponse (Maybe Text)

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

listOrganizationPortfolioAccessResponse_organizationNodes :: Lens' ListOrganizationPortfolioAccessResponse (Maybe [OrganizationNode])

Displays information about the organization nodes.

listOrganizationPortfolioAccessResponse_httpStatus :: Lens' ListOrganizationPortfolioAccessResponse Int

The response's http status code.