Amazonka.ServiceCatalog.SearchProducts

Description

Gets information about the products to which the caller has access.

Creating a Request

data SearchProducts

See: newSearchProducts smart constructor.

Constructors

SearchProducts'

Fields

filters :: Maybe (HashMap ProductViewFilterBy [Text])

The search filters. If no search filters are specified, the output includes all products to which the caller has access.

sortOrder :: Maybe SortOrder

The sort order. If no value is specified, the results are not sorted.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

pageToken :: Maybe Text

The page token for the next set of results. To retrieve the first set of results, use null.

pageSize :: Maybe Natural

The maximum number of items to return with this call.

sortBy :: Maybe ProductViewSortBy

The sort field. If no value is specified, the results are not sorted.

newSearchProducts :: SearchProducts

Create a value of SearchProducts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:SearchProducts', searchProducts_filters - The search filters. If no search filters are specified, the output includes all products to which the caller has access.

$sel:sortOrder:SearchProducts', searchProducts_sortOrder - The sort order. If no value is specified, the results are not sorted.

$sel:acceptLanguage:SearchProducts', searchProducts_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:pageToken:SearchProducts', searchProducts_pageToken - The page token for the next set of results. To retrieve the first set of results, use null.

$sel:pageSize:SearchProducts', searchProducts_pageSize - The maximum number of items to return with this call.

$sel:sortBy:SearchProducts', searchProducts_sortBy - The sort field. If no value is specified, the results are not sorted.

Request Lenses

searchProducts_filters :: Lens' SearchProducts (Maybe (HashMap ProductViewFilterBy [Text]))

The search filters. If no search filters are specified, the output includes all products to which the caller has access.

searchProducts_sortOrder :: Lens' SearchProducts (Maybe SortOrder)

The sort order. If no value is specified, the results are not sorted.

searchProducts_acceptLanguage :: Lens' SearchProducts (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

searchProducts_pageToken :: Lens' SearchProducts (Maybe Text)

The page token for the next set of results. To retrieve the first set of results, use null.

searchProducts_pageSize :: Lens' SearchProducts (Maybe Natural)

The maximum number of items to return with this call.

searchProducts_sortBy :: Lens' SearchProducts (Maybe ProductViewSortBy)

The sort field. If no value is specified, the results are not sorted.

Destructuring the Response

data SearchProductsResponse

See: newSearchProductsResponse smart constructor.

Constructors

SearchProductsResponse'

Fields

nextPageToken :: Maybe Text

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

productViewAggregations :: Maybe (HashMap Text [ProductViewAggregationValue])

The product view aggregations.

productViewSummaries :: Maybe [ProductViewSummary]

Information about the product views.

httpStatus :: Int

The response's http status code.

newSearchProductsResponse

Arguments

:: Int
$sel:httpStatus:SearchProductsResponse'

-> SearchProductsResponse

Create a value of SearchProductsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:SearchProductsResponse', searchProductsResponse_nextPageToken - The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

$sel:productViewAggregations:SearchProductsResponse', searchProductsResponse_productViewAggregations - The product view aggregations.

$sel:productViewSummaries:SearchProductsResponse', searchProductsResponse_productViewSummaries - Information about the product views.

$sel:httpStatus:SearchProductsResponse', searchProductsResponse_httpStatus - The response's http status code.

Response Lenses

searchProductsResponse_nextPageToken :: Lens' SearchProductsResponse (Maybe Text)

The page token to use to retrieve the next set of results. If there are no additional results, this value is null.

searchProductsResponse_productViewAggregations :: Lens' SearchProductsResponse (Maybe (HashMap Text [ProductViewAggregationValue]))

The product view aggregations.

searchProductsResponse_productViewSummaries :: Lens' SearchProductsResponse (Maybe [ProductViewSummary])

Information about the product views.

searchProductsResponse_httpStatus :: Lens' SearchProductsResponse Int

The response's http status code.