Amazonka.ServiceCatalog.Types.ListRecordHistorySearchFilter

Description

Documentation

data ListRecordHistorySearchFilter

The search filter to use when listing history records.

See: newListRecordHistorySearchFilter smart constructor.

Constructors

ListRecordHistorySearchFilter'

Fields

value :: Maybe Text

The filter value.

key :: Maybe Text

The filter key.

product - Filter results based on the specified product identifier.

provisionedproduct - Filter results based on the provisioned product identifier.

newListRecordHistorySearchFilter :: ListRecordHistorySearchFilter

Create a value of ListRecordHistorySearchFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:ListRecordHistorySearchFilter', listRecordHistorySearchFilter_value - The filter value.

$sel:key:ListRecordHistorySearchFilter', listRecordHistorySearchFilter_key - The filter key.

product - Filter results based on the specified product identifier.

provisionedproduct - Filter results based on the provisioned product identifier.

listRecordHistorySearchFilter_value :: Lens' ListRecordHistorySearchFilter (Maybe Text)

The filter value.

listRecordHistorySearchFilter_key :: Lens' ListRecordHistorySearchFilter (Maybe Text)

The filter key.

product - Filter results based on the specified product identifier.

provisionedproduct - Filter results based on the provisioned product identifier.