Amazonka.ServiceCatalog.Types.ParameterConstraints

Description

Documentation

data ParameterConstraints

The constraints that the administrator has put on the parameter.

See: newParameterConstraints smart constructor.

Constructors

ParameterConstraints'

Fields

maxValue :: Maybe Text

A numeric value that determines the largest numeric value you want to allow for Number types.

maxLength :: Maybe Text

An integer value that determines the largest number of characters you want to allow for String types.

constraintDescription :: Maybe Text

A string that explains a constraint when the constraint is violated. For example, without a constraint description, a parameter that has an allowed pattern of [A-Za-z0-9]+ displays the following error message when the user specifies an invalid value:

Malformed input-Parameter MyParameter must match pattern [A-Za-z0-9]+

By adding a constraint description, such as must only contain letters (uppercase and lowercase) and numbers, you can display the following customized error message:

Malformed input-Parameter MyParameter must only contain uppercase and lowercase letters and numbers.

minLength :: Maybe Text

An integer value that determines the smallest number of characters you want to allow for String types.

allowedPattern :: Maybe Text

A regular expression that represents the patterns that allow for String types. The pattern must match the entire parameter value provided.

allowedValues :: Maybe [Text]

The values that the administrator has allowed for the parameter.

minValue :: Maybe Text

A numeric value that determines the smallest numeric value you want to allow for Number types.

newParameterConstraints :: ParameterConstraints

Create a value of ParameterConstraints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxValue:ParameterConstraints', parameterConstraints_maxValue - A numeric value that determines the largest numeric value you want to allow for Number types.

$sel:maxLength:ParameterConstraints', parameterConstraints_maxLength - An integer value that determines the largest number of characters you want to allow for String types.

$sel:constraintDescription:ParameterConstraints', parameterConstraints_constraintDescription - A string that explains a constraint when the constraint is violated. For example, without a constraint description, a parameter that has an allowed pattern of [A-Za-z0-9]+ displays the following error message when the user specifies an invalid value:

Malformed input-Parameter MyParameter must match pattern [A-Za-z0-9]+

By adding a constraint description, such as must only contain letters (uppercase and lowercase) and numbers, you can display the following customized error message:

Malformed input-Parameter MyParameter must only contain uppercase and lowercase letters and numbers.

$sel:minLength:ParameterConstraints', parameterConstraints_minLength - An integer value that determines the smallest number of characters you want to allow for String types.

$sel:allowedPattern:ParameterConstraints', parameterConstraints_allowedPattern - A regular expression that represents the patterns that allow for String types. The pattern must match the entire parameter value provided.

$sel:allowedValues:ParameterConstraints', parameterConstraints_allowedValues - The values that the administrator has allowed for the parameter.

$sel:minValue:ParameterConstraints', parameterConstraints_minValue - A numeric value that determines the smallest numeric value you want to allow for Number types.

parameterConstraints_maxValue :: Lens' ParameterConstraints (Maybe Text)

A numeric value that determines the largest numeric value you want to allow for Number types.

parameterConstraints_maxLength :: Lens' ParameterConstraints (Maybe Text)

An integer value that determines the largest number of characters you want to allow for String types.

parameterConstraints_constraintDescription :: Lens' ParameterConstraints (Maybe Text)

A string that explains a constraint when the constraint is violated. For example, without a constraint description, a parameter that has an allowed pattern of [A-Za-z0-9]+ displays the following error message when the user specifies an invalid value:

Malformed input-Parameter MyParameter must match pattern [A-Za-z0-9]+

By adding a constraint description, such as must only contain letters (uppercase and lowercase) and numbers, you can display the following customized error message:

Malformed input-Parameter MyParameter must only contain uppercase and lowercase letters and numbers.

parameterConstraints_minLength :: Lens' ParameterConstraints (Maybe Text)

An integer value that determines the smallest number of characters you want to allow for String types.

parameterConstraints_allowedPattern :: Lens' ParameterConstraints (Maybe Text)

A regular expression that represents the patterns that allow for String types. The pattern must match the entire parameter value provided.

parameterConstraints_allowedValues :: Lens' ParameterConstraints (Maybe [Text])

The values that the administrator has allowed for the parameter.

parameterConstraints_minValue :: Lens' ParameterConstraints (Maybe Text)

A numeric value that determines the smallest numeric value you want to allow for Number types.