Amazonka.ServiceCatalog.Types.ProductViewSummary

Description

Documentation

data ProductViewSummary

Summary information about a product view.

See: newProductViewSummary smart constructor.

Constructors

ProductViewSummary'

Fields

owner :: Maybe Text

The owner of the product. Contact the product administrator for the significance of this value.

supportUrl :: Maybe Text

The URL information to obtain support for this Product.

shortDescription :: Maybe Text

Short description of the product.

hasDefaultPath :: Maybe Bool

Indicates whether the product has a default path. If the product does not have a default path, call ListLaunchPaths to disambiguate between paths. Otherwise, ListLaunchPaths is not required, and the output of ProductViewSummary can be used directly with DescribeProvisioningParameters.

distributor :: Maybe Text

The distributor of the product. Contact the product administrator for the significance of this value.

name :: Maybe Text

The name of the product.

id :: Maybe Text

The product view identifier.

type' :: Maybe ProductType

The product type. Contact the product administrator for the significance of this value. If this value is MARKETPLACE, the product was created by AWS Marketplace.

supportEmail :: Maybe Text

The email contact information to obtain support for this Product.

productId :: Maybe Text

The product identifier.

supportDescription :: Maybe Text

The description of the support for this Product.

newProductViewSummary :: ProductViewSummary

Create a value of ProductViewSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:owner:ProductViewSummary', productViewSummary_owner - The owner of the product. Contact the product administrator for the significance of this value.

$sel:supportUrl:ProductViewSummary', productViewSummary_supportUrl - The URL information to obtain support for this Product.

$sel:shortDescription:ProductViewSummary', productViewSummary_shortDescription - Short description of the product.

$sel:hasDefaultPath:ProductViewSummary', productViewSummary_hasDefaultPath - Indicates whether the product has a default path. If the product does not have a default path, call ListLaunchPaths to disambiguate between paths. Otherwise, ListLaunchPaths is not required, and the output of ProductViewSummary can be used directly with DescribeProvisioningParameters.

$sel:distributor:ProductViewSummary', productViewSummary_distributor - The distributor of the product. Contact the product administrator for the significance of this value.

$sel:name:ProductViewSummary', productViewSummary_name - The name of the product.

$sel:id:ProductViewSummary', productViewSummary_id - The product view identifier.

$sel:type':ProductViewSummary', productViewSummary_type - The product type. Contact the product administrator for the significance of this value. If this value is MARKETPLACE, the product was created by AWS Marketplace.

$sel:supportEmail:ProductViewSummary', productViewSummary_supportEmail - The email contact information to obtain support for this Product.

$sel:productId:ProductViewSummary', productViewSummary_productId - The product identifier.

$sel:supportDescription:ProductViewSummary', productViewSummary_supportDescription - The description of the support for this Product.

productViewSummary_owner :: Lens' ProductViewSummary (Maybe Text)

The owner of the product. Contact the product administrator for the significance of this value.

productViewSummary_supportUrl :: Lens' ProductViewSummary (Maybe Text)

The URL information to obtain support for this Product.

productViewSummary_shortDescription :: Lens' ProductViewSummary (Maybe Text)

Short description of the product.

productViewSummary_hasDefaultPath :: Lens' ProductViewSummary (Maybe Bool)

Indicates whether the product has a default path. If the product does not have a default path, call ListLaunchPaths to disambiguate between paths. Otherwise, ListLaunchPaths is not required, and the output of ProductViewSummary can be used directly with DescribeProvisioningParameters.

productViewSummary_distributor :: Lens' ProductViewSummary (Maybe Text)

The distributor of the product. Contact the product administrator for the significance of this value.

productViewSummary_name :: Lens' ProductViewSummary (Maybe Text)

The name of the product.

productViewSummary_id :: Lens' ProductViewSummary (Maybe Text)

The product view identifier.

productViewSummary_type :: Lens' ProductViewSummary (Maybe ProductType)

The product type. Contact the product administrator for the significance of this value. If this value is MARKETPLACE, the product was created by AWS Marketplace.

productViewSummary_supportEmail :: Lens' ProductViewSummary (Maybe Text)

The email contact information to obtain support for this Product.

productViewSummary_productId :: Lens' ProductViewSummary (Maybe Text)

The product identifier.

productViewSummary_supportDescription :: Lens' ProductViewSummary (Maybe Text)

The description of the support for this Product.