Amazonka.ServiceCatalog.Types.ProvisionedProductPlanDetails

Description

Documentation

data ProvisionedProductPlanDetails

Information about a plan.

See: newProvisionedProductPlanDetails smart constructor.

Constructors

ProvisionedProductPlanDetails'

Fields

status :: Maybe ProvisionedProductPlanStatus

The status.

provisionProductId :: Maybe Text

The product identifier.

provisioningArtifactId :: Maybe Text

The identifier of the provisioning artifact.

provisionProductName :: Maybe Text

The user-friendly name of the provisioned product.

createdTime :: Maybe POSIX

The UTC time stamp of the creation time.

notificationArns :: Maybe [Text]

Passed to CloudFormation. The SNS topic ARNs to which to publish stack-related events.

planId :: Maybe Text

The plan identifier.

planName :: Maybe Text

The name of the plan.

statusMessage :: Maybe Text

The status message.

updatedTime :: Maybe POSIX

The time when the plan was last updated.

pathId :: Maybe Text

The path identifier of the product. This value is optional if the product has a default path, and required if the product has more than one path. To list the paths for a product, use ListLaunchPaths.

provisioningParameters :: Maybe [UpdateProvisioningParameter]

Parameters specified by the administrator that are required for provisioning the product.

planType :: Maybe ProvisionedProductPlanType

The plan type.

productId :: Maybe Text

The product identifier.

tags :: Maybe [Tag]

One or more tags.

newProvisionedProductPlanDetails :: ProvisionedProductPlanDetails

Create a value of ProvisionedProductPlanDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ProvisionedProductPlanDetails', provisionedProductPlanDetails_status - The status.

$sel:provisionProductId:ProvisionedProductPlanDetails', provisionedProductPlanDetails_provisionProductId - The product identifier.

$sel:provisioningArtifactId:ProvisionedProductPlanDetails', provisionedProductPlanDetails_provisioningArtifactId - The identifier of the provisioning artifact.

$sel:provisionProductName:ProvisionedProductPlanDetails', provisionedProductPlanDetails_provisionProductName - The user-friendly name of the provisioned product.

$sel:createdTime:ProvisionedProductPlanDetails', provisionedProductPlanDetails_createdTime - The UTC time stamp of the creation time.

$sel:notificationArns:ProvisionedProductPlanDetails', provisionedProductPlanDetails_notificationArns - Passed to CloudFormation. The SNS topic ARNs to which to publish stack-related events.

$sel:planId:ProvisionedProductPlanDetails', provisionedProductPlanDetails_planId - The plan identifier.

$sel:planName:ProvisionedProductPlanDetails', provisionedProductPlanDetails_planName - The name of the plan.

$sel:statusMessage:ProvisionedProductPlanDetails', provisionedProductPlanDetails_statusMessage - The status message.

$sel:updatedTime:ProvisionedProductPlanDetails', provisionedProductPlanDetails_updatedTime - The time when the plan was last updated.

$sel:pathId:ProvisionedProductPlanDetails', provisionedProductPlanDetails_pathId - The path identifier of the product. This value is optional if the product has a default path, and required if the product has more than one path. To list the paths for a product, use ListLaunchPaths.

$sel:provisioningParameters:ProvisionedProductPlanDetails', provisionedProductPlanDetails_provisioningParameters - Parameters specified by the administrator that are required for provisioning the product.

$sel:planType:ProvisionedProductPlanDetails', provisionedProductPlanDetails_planType - The plan type.

$sel:productId:ProvisionedProductPlanDetails', provisionedProductPlanDetails_productId - The product identifier.

$sel:tags:ProvisionedProductPlanDetails', provisionedProductPlanDetails_tags - One or more tags.

provisionedProductPlanDetails_status :: Lens' ProvisionedProductPlanDetails (Maybe ProvisionedProductPlanStatus)

The status.

provisionedProductPlanDetails_provisionProductId :: Lens' ProvisionedProductPlanDetails (Maybe Text)

The product identifier.

provisionedProductPlanDetails_provisioningArtifactId :: Lens' ProvisionedProductPlanDetails (Maybe Text)

The identifier of the provisioning artifact.

provisionedProductPlanDetails_provisionProductName :: Lens' ProvisionedProductPlanDetails (Maybe Text)

The user-friendly name of the provisioned product.

provisionedProductPlanDetails_createdTime :: Lens' ProvisionedProductPlanDetails (Maybe UTCTime)

The UTC time stamp of the creation time.

provisionedProductPlanDetails_notificationArns :: Lens' ProvisionedProductPlanDetails (Maybe [Text])

Passed to CloudFormation. The SNS topic ARNs to which to publish stack-related events.

provisionedProductPlanDetails_planId :: Lens' ProvisionedProductPlanDetails (Maybe Text)

The plan identifier.

provisionedProductPlanDetails_planName :: Lens' ProvisionedProductPlanDetails (Maybe Text)

The name of the plan.

provisionedProductPlanDetails_statusMessage :: Lens' ProvisionedProductPlanDetails (Maybe Text)

The status message.

provisionedProductPlanDetails_updatedTime :: Lens' ProvisionedProductPlanDetails (Maybe UTCTime)

The time when the plan was last updated.

provisionedProductPlanDetails_pathId :: Lens' ProvisionedProductPlanDetails (Maybe Text)

The path identifier of the product. This value is optional if the product has a default path, and required if the product has more than one path. To list the paths for a product, use ListLaunchPaths.

provisionedProductPlanDetails_provisioningParameters :: Lens' ProvisionedProductPlanDetails (Maybe [UpdateProvisioningParameter])

Parameters specified by the administrator that are required for provisioning the product.

provisionedProductPlanDetails_planType :: Lens' ProvisionedProductPlanDetails (Maybe ProvisionedProductPlanType)

The plan type.

provisionedProductPlanDetails_productId :: Lens' ProvisionedProductPlanDetails (Maybe Text)

The product identifier.

provisionedProductPlanDetails_tags :: Lens' ProvisionedProductPlanDetails (Maybe [Tag])

One or more tags.