Amazonka.ServiceCatalog.Types.ProvisioningArtifactDetail

Description

Documentation

data ProvisioningArtifactDetail

Information about a provisioning artifact (also known as a version) for a product.

See: newProvisioningArtifactDetail smart constructor.

Constructors

ProvisioningArtifactDetail'

Fields

createdTime :: Maybe POSIX

The UTC time stamp of the creation time.

active :: Maybe Bool

Indicates whether the product version is active.

name :: Maybe Text

The name of the provisioning artifact.

id :: Maybe Text

The identifier of the provisioning artifact.

type' :: Maybe ProvisioningArtifactType

The type of provisioning artifact.

CLOUD_FORMATION_TEMPLATE - AWS CloudFormation template

MARKETPLACE_AMI - AWS Marketplace AMI

MARKETPLACE_CAR - AWS Marketplace Clusters and AWS Resources

guidance :: Maybe ProvisioningArtifactGuidance

Information set by the administrator to provide guidance to end users about which provisioning artifacts to use.

description :: Maybe Text

The description of the provisioning artifact.

newProvisioningArtifactDetail :: ProvisioningArtifactDetail

Create a value of ProvisioningArtifactDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdTime:ProvisioningArtifactDetail', provisioningArtifactDetail_createdTime - The UTC time stamp of the creation time.

$sel:active:ProvisioningArtifactDetail', provisioningArtifactDetail_active - Indicates whether the product version is active.

$sel:name:ProvisioningArtifactDetail', provisioningArtifactDetail_name - The name of the provisioning artifact.

$sel:id:ProvisioningArtifactDetail', provisioningArtifactDetail_id - The identifier of the provisioning artifact.

$sel:type':ProvisioningArtifactDetail', provisioningArtifactDetail_type - The type of provisioning artifact.

CLOUD_FORMATION_TEMPLATE - AWS CloudFormation template

MARKETPLACE_AMI - AWS Marketplace AMI

MARKETPLACE_CAR - AWS Marketplace Clusters and AWS Resources

$sel:guidance:ProvisioningArtifactDetail', provisioningArtifactDetail_guidance - Information set by the administrator to provide guidance to end users about which provisioning artifacts to use.

$sel:description:ProvisioningArtifactDetail', provisioningArtifactDetail_description - The description of the provisioning artifact.

provisioningArtifactDetail_createdTime :: Lens' ProvisioningArtifactDetail (Maybe UTCTime)

The UTC time stamp of the creation time.

provisioningArtifactDetail_active :: Lens' ProvisioningArtifactDetail (Maybe Bool)

Indicates whether the product version is active.

provisioningArtifactDetail_name :: Lens' ProvisioningArtifactDetail (Maybe Text)

The name of the provisioning artifact.

provisioningArtifactDetail_id :: Lens' ProvisioningArtifactDetail (Maybe Text)

The identifier of the provisioning artifact.

provisioningArtifactDetail_type :: Lens' ProvisioningArtifactDetail (Maybe ProvisioningArtifactType)

The type of provisioning artifact.

CLOUD_FORMATION_TEMPLATE - AWS CloudFormation template

MARKETPLACE_AMI - AWS Marketplace AMI

MARKETPLACE_CAR - AWS Marketplace Clusters and AWS Resources

provisioningArtifactDetail_guidance :: Lens' ProvisioningArtifactDetail (Maybe ProvisioningArtifactGuidance)

Information set by the administrator to provide guidance to end users about which provisioning artifacts to use.

provisioningArtifactDetail_description :: Lens' ProvisioningArtifactDetail (Maybe Text)

The description of the provisioning artifact.