Amazonka.ServiceCatalog.Types.ProvisioningArtifactParameter

Description

Documentation

data ProvisioningArtifactParameter

Information about a parameter used to provision a product.

See: newProvisioningArtifactParameter smart constructor.

Constructors

ProvisioningArtifactParameter'

Fields

isNoEcho :: Maybe Bool

If this value is true, the value for this parameter is obfuscated from view when the parameter is retrieved. This parameter is used to hide sensitive information.

parameterKey :: Maybe Text

The parameter key.

parameterType :: Maybe Text

The parameter type.

parameterConstraints :: Maybe ParameterConstraints

Constraints that the administrator has put on a parameter.

defaultValue :: Maybe Text

The default value.

description :: Maybe Text

The description of the parameter.

newProvisioningArtifactParameter :: ProvisioningArtifactParameter

Create a value of ProvisioningArtifactParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isNoEcho:ProvisioningArtifactParameter', provisioningArtifactParameter_isNoEcho - If this value is true, the value for this parameter is obfuscated from view when the parameter is retrieved. This parameter is used to hide sensitive information.

$sel:parameterKey:ProvisioningArtifactParameter', provisioningArtifactParameter_parameterKey - The parameter key.

$sel:parameterType:ProvisioningArtifactParameter', provisioningArtifactParameter_parameterType - The parameter type.

$sel:parameterConstraints:ProvisioningArtifactParameter', provisioningArtifactParameter_parameterConstraints - Constraints that the administrator has put on a parameter.

$sel:defaultValue:ProvisioningArtifactParameter', provisioningArtifactParameter_defaultValue - The default value.

$sel:description:ProvisioningArtifactParameter', provisioningArtifactParameter_description - The description of the parameter.

provisioningArtifactParameter_isNoEcho :: Lens' ProvisioningArtifactParameter (Maybe Bool)

If this value is true, the value for this parameter is obfuscated from view when the parameter is retrieved. This parameter is used to hide sensitive information.

provisioningArtifactParameter_parameterKey :: Lens' ProvisioningArtifactParameter (Maybe Text)

The parameter key.

provisioningArtifactParameter_parameterType :: Lens' ProvisioningArtifactParameter (Maybe Text)

The parameter type.

provisioningArtifactParameter_parameterConstraints :: Lens' ProvisioningArtifactParameter (Maybe ParameterConstraints)

Constraints that the administrator has put on a parameter.

provisioningArtifactParameter_defaultValue :: Lens' ProvisioningArtifactParameter (Maybe Text)

The default value.

provisioningArtifactParameter_description :: Lens' ProvisioningArtifactParameter (Maybe Text)

The description of the parameter.