Amazonka.ServiceCatalog.UpdatePortfolio

Description

Updates the specified portfolio.

You cannot update a product that was shared with you.

Creating a Request

data UpdatePortfolio

See: newUpdatePortfolio smart constructor.

Constructors

UpdatePortfolio'

Fields

removeTags :: Maybe [Text]

The tags to remove.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

displayName :: Maybe Text

The name to use for display purposes.

addTags :: Maybe [Tag]

The tags to add.

description :: Maybe Text

The updated description of the portfolio.

providerName :: Maybe Text

The updated name of the portfolio provider.

id :: Text

The portfolio identifier.

newUpdatePortfolio

Arguments

:: Text
$sel:id:UpdatePortfolio'

-> UpdatePortfolio

Create a value of UpdatePortfolio with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:removeTags:UpdatePortfolio', updatePortfolio_removeTags - The tags to remove.

$sel:acceptLanguage:UpdatePortfolio', updatePortfolio_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:displayName:UpdatePortfolio', updatePortfolio_displayName - The name to use for display purposes.

$sel:addTags:UpdatePortfolio', updatePortfolio_addTags - The tags to add.

$sel:description:UpdatePortfolio', updatePortfolio_description - The updated description of the portfolio.

$sel:providerName:UpdatePortfolio', updatePortfolio_providerName - The updated name of the portfolio provider.

$sel:id:UpdatePortfolio', updatePortfolio_id - The portfolio identifier.

Request Lenses

updatePortfolio_removeTags :: Lens' UpdatePortfolio (Maybe [Text])

The tags to remove.

updatePortfolio_acceptLanguage :: Lens' UpdatePortfolio (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

updatePortfolio_displayName :: Lens' UpdatePortfolio (Maybe Text)

The name to use for display purposes.

updatePortfolio_addTags :: Lens' UpdatePortfolio (Maybe [Tag])

The tags to add.

updatePortfolio_description :: Lens' UpdatePortfolio (Maybe Text)

The updated description of the portfolio.

updatePortfolio_providerName :: Lens' UpdatePortfolio (Maybe Text)

The updated name of the portfolio provider.

updatePortfolio_id :: Lens' UpdatePortfolio Text

The portfolio identifier.

Destructuring the Response

data UpdatePortfolioResponse

See: newUpdatePortfolioResponse smart constructor.

Constructors

UpdatePortfolioResponse'

Fields

portfolioDetail :: Maybe PortfolioDetail

Information about the portfolio.

tags :: Maybe [Tag]

Information about the tags associated with the portfolio.

httpStatus :: Int

The response's http status code.

newUpdatePortfolioResponse

Arguments

:: Int
$sel:httpStatus:UpdatePortfolioResponse'

-> UpdatePortfolioResponse

Create a value of UpdatePortfolioResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portfolioDetail:UpdatePortfolioResponse', updatePortfolioResponse_portfolioDetail - Information about the portfolio.

$sel:tags:UpdatePortfolioResponse', updatePortfolioResponse_tags - Information about the tags associated with the portfolio.

$sel:httpStatus:UpdatePortfolioResponse', updatePortfolioResponse_httpStatus - The response's http status code.

Response Lenses

updatePortfolioResponse_portfolioDetail :: Lens' UpdatePortfolioResponse (Maybe PortfolioDetail)

Information about the portfolio.

updatePortfolioResponse_tags :: Lens' UpdatePortfolioResponse (Maybe [Tag])

Information about the tags associated with the portfolio.

updatePortfolioResponse_httpStatus :: Lens' UpdatePortfolioResponse Int

The response's http status code.