Amazonka.ServiceCatalog.UpdateProvisioningArtifact

Description

Updates the specified provisioning artifact (also known as a version) for the specified product.

You cannot update a provisioning artifact for a product that was shared with you.

Creating a Request

data UpdateProvisioningArtifact

See: newUpdateProvisioningArtifact smart constructor.

Constructors

UpdateProvisioningArtifact'

Fields

active :: Maybe Bool

Indicates whether the product version is active.

Inactive provisioning artifacts are invisible to end users. End users cannot launch or update a provisioned product from an inactive provisioning artifact.

name :: Maybe Text

The updated name of the provisioning artifact.

acceptLanguage :: Maybe Text

The language code.

en - English (default)

jp - Japanese

zh - Chinese

guidance :: Maybe ProvisioningArtifactGuidance

Information set by the administrator to provide guidance to end users about which provisioning artifacts to use.

The DEFAULT value indicates that the product version is active.

The administrator can set the guidance to DEPRECATED to inform users that the product version is deprecated. Users are able to make updates to a provisioned product of a deprecated version but cannot launch new provisioned products using a deprecated version.

description :: Maybe Text

The updated description of the provisioning artifact.

productId :: Text

The product identifier.

provisioningArtifactId :: Text

The identifier of the provisioning artifact.

newUpdateProvisioningArtifact

Arguments

:: Text
$sel:productId:UpdateProvisioningArtifact'

-> Text
$sel:provisioningArtifactId:UpdateProvisioningArtifact'

-> UpdateProvisioningArtifact

Create a value of UpdateProvisioningArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:active:UpdateProvisioningArtifact', updateProvisioningArtifact_active - Indicates whether the product version is active.

Inactive provisioning artifacts are invisible to end users. End users cannot launch or update a provisioned product from an inactive provisioning artifact.

$sel:name:UpdateProvisioningArtifact', updateProvisioningArtifact_name - The updated name of the provisioning artifact.

$sel:acceptLanguage:UpdateProvisioningArtifact', updateProvisioningArtifact_acceptLanguage - The language code.

en - English (default)

jp - Japanese

zh - Chinese

$sel:guidance:UpdateProvisioningArtifact', updateProvisioningArtifact_guidance - Information set by the administrator to provide guidance to end users about which provisioning artifacts to use.

The DEFAULT value indicates that the product version is active.

The administrator can set the guidance to DEPRECATED to inform users that the product version is deprecated. Users are able to make updates to a provisioned product of a deprecated version but cannot launch new provisioned products using a deprecated version.

$sel:description:UpdateProvisioningArtifact', updateProvisioningArtifact_description - The updated description of the provisioning artifact.

$sel:productId:UpdateProvisioningArtifact', updateProvisioningArtifact_productId - The product identifier.

$sel:provisioningArtifactId:UpdateProvisioningArtifact', updateProvisioningArtifact_provisioningArtifactId - The identifier of the provisioning artifact.

Request Lenses

updateProvisioningArtifact_active :: Lens' UpdateProvisioningArtifact (Maybe Bool)

Indicates whether the product version is active.

Inactive provisioning artifacts are invisible to end users. End users cannot launch or update a provisioned product from an inactive provisioning artifact.

updateProvisioningArtifact_name :: Lens' UpdateProvisioningArtifact (Maybe Text)

The updated name of the provisioning artifact.

updateProvisioningArtifact_acceptLanguage :: Lens' UpdateProvisioningArtifact (Maybe Text)

The language code.

en - English (default)

jp - Japanese

zh - Chinese

updateProvisioningArtifact_guidance :: Lens' UpdateProvisioningArtifact (Maybe ProvisioningArtifactGuidance)

Information set by the administrator to provide guidance to end users about which provisioning artifacts to use.

The DEFAULT value indicates that the product version is active.

The administrator can set the guidance to DEPRECATED to inform users that the product version is deprecated. Users are able to make updates to a provisioned product of a deprecated version but cannot launch new provisioned products using a deprecated version.

updateProvisioningArtifact_description :: Lens' UpdateProvisioningArtifact (Maybe Text)

The updated description of the provisioning artifact.

updateProvisioningArtifact_productId :: Lens' UpdateProvisioningArtifact Text

The product identifier.

updateProvisioningArtifact_provisioningArtifactId :: Lens' UpdateProvisioningArtifact Text

The identifier of the provisioning artifact.

Destructuring the Response

data UpdateProvisioningArtifactResponse

See: newUpdateProvisioningArtifactResponse smart constructor.

Constructors

UpdateProvisioningArtifactResponse'

Fields

status :: Maybe RequestStatus

The status of the current request.

info :: Maybe (HashMap Text Text)

The URL of the CloudFormation template in Amazon S3.

provisioningArtifactDetail :: Maybe ProvisioningArtifactDetail

Information about the provisioning artifact.

httpStatus :: Int

The response's http status code.

newUpdateProvisioningArtifactResponse

Arguments

:: Int
$sel:httpStatus:UpdateProvisioningArtifactResponse'

-> UpdateProvisioningArtifactResponse

Create a value of UpdateProvisioningArtifactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateProvisioningArtifactResponse', updateProvisioningArtifactResponse_status - The status of the current request.

$sel:info:UpdateProvisioningArtifactResponse', updateProvisioningArtifactResponse_info - The URL of the CloudFormation template in Amazon S3.

$sel:provisioningArtifactDetail:UpdateProvisioningArtifactResponse', updateProvisioningArtifactResponse_provisioningArtifactDetail - Information about the provisioning artifact.

$sel:httpStatus:UpdateProvisioningArtifactResponse', updateProvisioningArtifactResponse_httpStatus - The response's http status code.

Response Lenses

updateProvisioningArtifactResponse_status :: Lens' UpdateProvisioningArtifactResponse (Maybe RequestStatus)

The status of the current request.

updateProvisioningArtifactResponse_info :: Lens' UpdateProvisioningArtifactResponse (Maybe (HashMap Text Text))

The URL of the CloudFormation template in Amazon S3.

updateProvisioningArtifactResponse_provisioningArtifactDetail :: Lens' UpdateProvisioningArtifactResponse (Maybe ProvisioningArtifactDetail)

Information about the provisioning artifact.

updateProvisioningArtifactResponse_httpStatus :: Lens' UpdateProvisioningArtifactResponse Int

The response's http status code.