Amazonka.ServiceCatalog.UpdateTagOption

Description

Updates the specified TagOption.

Creating a Request

data UpdateTagOption

See: newUpdateTagOption smart constructor.

Constructors

UpdateTagOption'

Fields

value :: Maybe Text

The updated value.

active :: Maybe Bool

The updated active state.

id :: Text

The TagOption identifier.

newUpdateTagOption

Arguments

:: Text
$sel:id:UpdateTagOption'

-> UpdateTagOption

Create a value of UpdateTagOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:UpdateTagOption', updateTagOption_value - The updated value.

$sel:active:UpdateTagOption', updateTagOption_active - The updated active state.

$sel:id:UpdateTagOption', updateTagOption_id - The TagOption identifier.

Request Lenses

updateTagOption_value :: Lens' UpdateTagOption (Maybe Text)

The updated value.

updateTagOption_active :: Lens' UpdateTagOption (Maybe Bool)

The updated active state.

updateTagOption_id :: Lens' UpdateTagOption Text

The TagOption identifier.

Destructuring the Response

data UpdateTagOptionResponse

See: newUpdateTagOptionResponse smart constructor.

Constructors

UpdateTagOptionResponse'

Fields

tagOptionDetail :: Maybe TagOptionDetail

Information about the TagOption.

httpStatus :: Int

The response's http status code.

newUpdateTagOptionResponse

Arguments

:: Int
$sel:httpStatus:UpdateTagOptionResponse'

-> UpdateTagOptionResponse

Create a value of UpdateTagOptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagOptionDetail:UpdateTagOptionResponse', updateTagOptionResponse_tagOptionDetail - Information about the TagOption.

$sel:httpStatus:UpdateTagOptionResponse', updateTagOptionResponse_httpStatus - The response's http status code.

Response Lenses

updateTagOptionResponse_tagOptionDetail :: Lens' UpdateTagOptionResponse (Maybe TagOptionDetail)

Information about the TagOption.

updateTagOptionResponse_httpStatus :: Lens' UpdateTagOptionResponse Int

The response's http status code.