Amazonka.ServiceCatalogAppRegistry

Description

Derived from API version 2020-06-24 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Web Services Service Catalog AppRegistry enables organizations to understand the application context of their Amazon Web Services resources. AppRegistry provides a repository of your applications, their resources, and the application metadata that you use within your enterprise.

Service Configuration

defaultService :: Service

API version 2020-06-24 of the Amazon Service Catalog App Registry SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ServiceCatalogAppRegistry.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The request has invalid or missing parameters.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

There was a conflict when processing the request (for example, a resource with the given name already exists within the account).

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of resources per account has been reached.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The service is experiencing internal problems.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

AssociateAttributeGroup

data AssociateAttributeGroup

See: newAssociateAttributeGroup smart constructor.

Constructors

AssociateAttributeGroup' Text Text

newAssociateAttributeGroup

Arguments

:: Text
$sel:application:AssociateAttributeGroup'

-> Text
$sel:attributeGroup:AssociateAttributeGroup'

-> AssociateAttributeGroup

Create a value of AssociateAttributeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:AssociateAttributeGroup', associateAttributeGroup_application - The name or ID of the application.

$sel:attributeGroup:AssociateAttributeGroup', associateAttributeGroup_attributeGroup - The name or ID of the attribute group that holds the attributes to describe the application.

data AssociateAttributeGroupResponse

See: newAssociateAttributeGroupResponse smart constructor.

Constructors

AssociateAttributeGroupResponse' (Maybe Text) (Maybe Text) Int

newAssociateAttributeGroupResponse

Arguments

:: Int
$sel:httpStatus:AssociateAttributeGroupResponse'

-> AssociateAttributeGroupResponse

Create a value of AssociateAttributeGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationArn:AssociateAttributeGroupResponse', associateAttributeGroupResponse_applicationArn - The Amazon resource name (ARN) of the application that was augmented with attributes.

$sel:attributeGroupArn:AssociateAttributeGroupResponse', associateAttributeGroupResponse_attributeGroupArn - The Amazon resource name (ARN) of the attribute group that contains the application's new attributes.

$sel:httpStatus:AssociateAttributeGroupResponse', associateAttributeGroupResponse_httpStatus - The response's http status code.

ListAttributeGroups (Paginated)

data ListAttributeGroups

See: newListAttributeGroups smart constructor.

Constructors

ListAttributeGroups' (Maybe Text) (Maybe Natural)

newListAttributeGroups :: ListAttributeGroups

Create a value of ListAttributeGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAttributeGroups', listAttributeGroups_nextToken - The token to use to get the next page of results after a previous API call.

$sel:maxResults:ListAttributeGroups', listAttributeGroups_maxResults - The upper bound of the number of results to return (cannot exceed 25). If this parameter is omitted, it defaults to 25. This value is optional.

data ListAttributeGroupsResponse

See: newListAttributeGroupsResponse smart constructor.

Constructors

ListAttributeGroupsResponse' (Maybe [AttributeGroupSummary]) (Maybe Text) Int

newListAttributeGroupsResponse

Arguments

:: Int
$sel:httpStatus:ListAttributeGroupsResponse'

-> ListAttributeGroupsResponse

Create a value of ListAttributeGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeGroups:ListAttributeGroupsResponse', listAttributeGroupsResponse_attributeGroups - This list of attribute groups.

$sel:nextToken:ListAttributeGroups', listAttributeGroupsResponse_nextToken - The token to use to get the next page of results after a previous API call.

$sel:httpStatus:ListAttributeGroupsResponse', listAttributeGroupsResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon resource name (ARN) that specifies the resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The tags on the resource.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

SyncResource

data SyncResource

See: newSyncResource smart constructor.

Constructors

SyncResource' ResourceType Text

newSyncResource

Arguments

:: ResourceType
$sel:resourceType:SyncResource'

-> Text
$sel:resource:SyncResource'

-> SyncResource

Create a value of SyncResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:SyncResource', syncResource_resourceType - The type of resource of which the application will be associated.

$sel:resource:SyncResource', syncResource_resource - An entity you can work with and specify with a name or ID. Examples include an Amazon EC2 instance, an Amazon Web Services CloudFormation stack, or an Amazon S3 bucket.

data SyncResourceResponse

See: newSyncResourceResponse smart constructor.

Constructors

SyncResourceResponse' (Maybe Text) (Maybe SyncAction) (Maybe Text) Int

newSyncResourceResponse

Arguments

:: Int
$sel:httpStatus:SyncResourceResponse'

-> SyncResourceResponse

Create a value of SyncResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationArn:SyncResourceResponse', syncResourceResponse_applicationArn - The Amazon resource name (ARN) that specifies the application.

$sel:actionTaken:SyncResourceResponse', syncResourceResponse_actionTaken - The results of the output if an application is associated with an ARN value, which could be syncStarted or None.

$sel:resourceArn:SyncResourceResponse', syncResourceResponse_resourceArn - The Amazon resource name (ARN) that specifies the resource.

$sel:httpStatus:SyncResourceResponse', syncResourceResponse_httpStatus - The response's http status code.

DeleteApplication

data DeleteApplication

See: newDeleteApplication smart constructor.

Constructors

DeleteApplication' Text

newDeleteApplication

Arguments

:: Text
$sel:application:DeleteApplication'

-> DeleteApplication

Create a value of DeleteApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:DeleteApplication', deleteApplication_application - The name or ID of the application.

data DeleteApplicationResponse

See: newDeleteApplicationResponse smart constructor.

Constructors

DeleteApplicationResponse' (Maybe ApplicationSummary) Int

newDeleteApplicationResponse

Arguments

:: Int
$sel:httpStatus:DeleteApplicationResponse'

-> DeleteApplicationResponse

Create a value of DeleteApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:DeleteApplication', deleteApplicationResponse_application - Information about the deleted application.

$sel:httpStatus:DeleteApplicationResponse', deleteApplicationResponse_httpStatus - The response's http status code.

UpdateApplication

data UpdateApplication

See: newUpdateApplication smart constructor.

Constructors

UpdateApplication' (Maybe Text) (Maybe Text) Text

newUpdateApplication

Arguments

:: Text
$sel:application:UpdateApplication'

-> UpdateApplication

Create a value of UpdateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateApplication', updateApplication_name - The new name of the application. The name must be unique in the region in which you are updating the application.

$sel:description:UpdateApplication', updateApplication_description - The new description of the application.

$sel:application:UpdateApplication', updateApplication_application - The name or ID of the application that will be updated.

data UpdateApplicationResponse

See: newUpdateApplicationResponse smart constructor.

Constructors

UpdateApplicationResponse' (Maybe Application) Int

newUpdateApplicationResponse

Arguments

:: Int
$sel:httpStatus:UpdateApplicationResponse'

-> UpdateApplicationResponse

Create a value of UpdateApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:UpdateApplication', updateApplicationResponse_application - The updated information of the application.

$sel:httpStatus:UpdateApplicationResponse', updateApplicationResponse_httpStatus - The response's http status code.

AssociateResource

data AssociateResource

See: newAssociateResource smart constructor.

Constructors

AssociateResource' Text ResourceType Text

newAssociateResource

Arguments

:: Text
$sel:application:AssociateResource'

-> ResourceType
$sel:resourceType:AssociateResource'

-> Text
$sel:resource:AssociateResource'

-> AssociateResource

Create a value of AssociateResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:AssociateResource', associateResource_application - The name or ID of the application.

$sel:resourceType:AssociateResource', associateResource_resourceType - The type of resource of which the application will be associated.

$sel:resource:AssociateResource', associateResource_resource - The name or ID of the resource of which the application will be associated.

data AssociateResourceResponse

See: newAssociateResourceResponse smart constructor.

Constructors

AssociateResourceResponse' (Maybe Text) (Maybe Text) Int

newAssociateResourceResponse

Arguments

:: Int
$sel:httpStatus:AssociateResourceResponse'

-> AssociateResourceResponse

Create a value of AssociateResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationArn:AssociateResourceResponse', associateResourceResponse_applicationArn - The Amazon resource name (ARN) of the application that was augmented with attributes.

$sel:resourceArn:AssociateResourceResponse', associateResourceResponse_resourceArn - The Amazon resource name (ARN) that specifies the resource.

$sel:httpStatus:AssociateResourceResponse', associateResourceResponse_httpStatus - The response's http status code.

CreateApplication

data CreateApplication

See: newCreateApplication smart constructor.

Constructors

CreateApplication' (Maybe Text) (Maybe (HashMap Text Text)) Text Text

newCreateApplication

Arguments

:: Text
$sel:name:CreateApplication'

-> Text
$sel:clientToken:CreateApplication'

-> CreateApplication

Create a value of CreateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateApplication', createApplication_description - The description of the application.

$sel:tags:CreateApplication', createApplication_tags - Key-value pairs you can use to associate with the application.

$sel:name:CreateApplication', createApplication_name - The name of the application. The name must be unique in the region in which you are creating the application.

$sel:clientToken:CreateApplication', createApplication_clientToken - A unique identifier that you provide to ensure idempotency. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails.

data CreateApplicationResponse

See: newCreateApplicationResponse smart constructor.

Constructors

CreateApplicationResponse' (Maybe Application) Int

newCreateApplicationResponse

Arguments

:: Int
$sel:httpStatus:CreateApplicationResponse'

-> CreateApplicationResponse

Create a value of CreateApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:CreateApplicationResponse', createApplicationResponse_application - Information about the application.

$sel:httpStatus:CreateApplicationResponse', createApplicationResponse_httpStatus - The response's http status code.

DisassociateAttributeGroup

data DisassociateAttributeGroup

See: newDisassociateAttributeGroup smart constructor.

Constructors

DisassociateAttributeGroup' Text Text

newDisassociateAttributeGroup

Arguments

:: Text
$sel:application:DisassociateAttributeGroup'

-> Text
$sel:attributeGroup:DisassociateAttributeGroup'

-> DisassociateAttributeGroup

Create a value of DisassociateAttributeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:DisassociateAttributeGroup', disassociateAttributeGroup_application - The name or ID of the application.

$sel:attributeGroup:DisassociateAttributeGroup', disassociateAttributeGroup_attributeGroup - The name or ID of the attribute group that holds the attributes to describe the application.

data DisassociateAttributeGroupResponse

See: newDisassociateAttributeGroupResponse smart constructor.

Constructors

DisassociateAttributeGroupResponse' (Maybe Text) (Maybe Text) Int

newDisassociateAttributeGroupResponse

Arguments

:: Int
$sel:httpStatus:DisassociateAttributeGroupResponse'

-> DisassociateAttributeGroupResponse

Create a value of DisassociateAttributeGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationArn:DisassociateAttributeGroupResponse', disassociateAttributeGroupResponse_applicationArn - The Amazon resource name (ARN) that specifies the application.

$sel:attributeGroupArn:DisassociateAttributeGroupResponse', disassociateAttributeGroupResponse_attributeGroupArn - The Amazon resource name (ARN) that specifies the attribute group.

$sel:httpStatus:DisassociateAttributeGroupResponse', disassociateAttributeGroupResponse_httpStatus - The response's http status code.

GetApplication

data GetApplication

See: newGetApplication smart constructor.

Constructors

GetApplication' Text

newGetApplication

Arguments

:: Text
$sel:application:GetApplication'

-> GetApplication

Create a value of GetApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:GetApplication', getApplication_application - The name or ID of the application.

data GetApplicationResponse

See: newGetApplicationResponse smart constructor.

Constructors

GetApplicationResponse' (Maybe POSIX) (Maybe Text) (Maybe Integrations) (Maybe Natural) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe (HashMap Text Text)) Int

newGetApplicationResponse

Arguments

:: Int
$sel:httpStatus:GetApplicationResponse'

-> GetApplicationResponse

Create a value of GetApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:GetApplicationResponse', getApplicationResponse_creationTime - The ISO-8601 formatted timestamp of the moment when the application was created.

$sel:arn:GetApplicationResponse', getApplicationResponse_arn - The Amazon resource name (ARN) that specifies the application across services.

$sel:integrations:GetApplicationResponse', getApplicationResponse_integrations - The information about the integration of the application with other services, such as Resource Groups.

$sel:associatedResourceCount:GetApplicationResponse', getApplicationResponse_associatedResourceCount - The number of top-level resources that were registered as part of this application.

$sel:name:GetApplicationResponse', getApplicationResponse_name - The name of the application. The name must be unique in the region in which you are creating the application.

$sel:id:GetApplicationResponse', getApplicationResponse_id - The identifier of the application.

$sel:lastUpdateTime:GetApplicationResponse', getApplicationResponse_lastUpdateTime - The ISO-8601 formatted timestamp of the moment when the application was last updated.

$sel:description:GetApplicationResponse', getApplicationResponse_description - The description of the application.

$sel:tags:GetApplicationResponse', getApplicationResponse_tags - Key-value pairs associated with the application.

$sel:httpStatus:GetApplicationResponse', getApplicationResponse_httpStatus - The response's http status code.

GetAssociatedResource

data GetAssociatedResource

See: newGetAssociatedResource smart constructor.

Constructors

GetAssociatedResource' Text ResourceType Text

newGetAssociatedResource

Arguments

:: Text
$sel:application:GetAssociatedResource'

-> ResourceType
$sel:resourceType:GetAssociatedResource'

-> Text
$sel:resource:GetAssociatedResource'

-> GetAssociatedResource

Create a value of GetAssociatedResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:GetAssociatedResource', getAssociatedResource_application - The name or ID of the application.

$sel:resourceType:GetAssociatedResource', getAssociatedResource_resourceType - The type of resource associated with the application.

$sel:resource:GetAssociatedResource', getAssociatedResource_resource - The name or ID of the resource associated with the application.

data GetAssociatedResourceResponse

See: newGetAssociatedResourceResponse smart constructor.

Constructors

GetAssociatedResourceResponse' (Maybe Resource) Int

newGetAssociatedResourceResponse

Arguments

:: Int
$sel:httpStatus:GetAssociatedResourceResponse'

-> GetAssociatedResourceResponse

Create a value of GetAssociatedResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resource:GetAssociatedResource', getAssociatedResourceResponse_resource - The resource associated with the application.

$sel:httpStatus:GetAssociatedResourceResponse', getAssociatedResourceResponse_httpStatus - The response's http status code.

CreateAttributeGroup

data CreateAttributeGroup

See: newCreateAttributeGroup smart constructor.

Constructors

CreateAttributeGroup' (Maybe Text) (Maybe (HashMap Text Text)) Text Text Text

newCreateAttributeGroup

Arguments

:: Text
$sel:name:CreateAttributeGroup'

-> Text
$sel:attributes:CreateAttributeGroup'

-> Text
$sel:clientToken:CreateAttributeGroup'

-> CreateAttributeGroup

Create a value of CreateAttributeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateAttributeGroup', createAttributeGroup_description - The description of the attribute group that the user provides.

$sel:tags:CreateAttributeGroup', createAttributeGroup_tags - Key-value pairs you can use to associate with the attribute group.

$sel:name:CreateAttributeGroup', createAttributeGroup_name - The name of the attribute group.

$sel:attributes:CreateAttributeGroup', createAttributeGroup_attributes - A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

$sel:clientToken:CreateAttributeGroup', createAttributeGroup_clientToken - A unique identifier that you provide to ensure idempotency. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails.

data CreateAttributeGroupResponse

See: newCreateAttributeGroupResponse smart constructor.

Constructors

CreateAttributeGroupResponse' (Maybe AttributeGroup) Int

newCreateAttributeGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateAttributeGroupResponse'

-> CreateAttributeGroupResponse

Create a value of CreateAttributeGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeGroup:CreateAttributeGroupResponse', createAttributeGroupResponse_attributeGroup - Information about the attribute group.

$sel:httpStatus:CreateAttributeGroupResponse', createAttributeGroupResponse_httpStatus - The response's http status code.

DeleteAttributeGroup

data DeleteAttributeGroup

See: newDeleteAttributeGroup smart constructor.

Constructors

DeleteAttributeGroup' Text

newDeleteAttributeGroup

Arguments

:: Text
$sel:attributeGroup:DeleteAttributeGroup'

-> DeleteAttributeGroup

Create a value of DeleteAttributeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeGroup:DeleteAttributeGroup', deleteAttributeGroup_attributeGroup - The name or ID of the attribute group that holds the attributes to describe the application.

data DeleteAttributeGroupResponse

See: newDeleteAttributeGroupResponse smart constructor.

Constructors

DeleteAttributeGroupResponse' (Maybe AttributeGroupSummary) Int

newDeleteAttributeGroupResponse

Arguments

:: Int
$sel:httpStatus:DeleteAttributeGroupResponse'

-> DeleteAttributeGroupResponse

Create a value of DeleteAttributeGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeGroup:DeleteAttributeGroup', deleteAttributeGroupResponse_attributeGroup - Information about the deleted attribute group.

$sel:httpStatus:DeleteAttributeGroupResponse', deleteAttributeGroupResponse_httpStatus - The response's http status code.

UpdateAttributeGroup

data UpdateAttributeGroup

See: newUpdateAttributeGroup smart constructor.

Constructors

UpdateAttributeGroup' (Maybe Text) (Maybe Text) (Maybe Text) Text

newUpdateAttributeGroup

Arguments

:: Text
$sel:attributeGroup:UpdateAttributeGroup'

-> UpdateAttributeGroup

Create a value of UpdateAttributeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateAttributeGroup', updateAttributeGroup_name - The new name of the attribute group. The name must be unique in the region in which you are updating the attribute group.

$sel:attributes:UpdateAttributeGroup', updateAttributeGroup_attributes - A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

$sel:description:UpdateAttributeGroup', updateAttributeGroup_description - The description of the attribute group that the user provides.

$sel:attributeGroup:UpdateAttributeGroup', updateAttributeGroup_attributeGroup - The name or ID of the attribute group that holds the attributes to describe the application.

data UpdateAttributeGroupResponse

See: newUpdateAttributeGroupResponse smart constructor.

Constructors

UpdateAttributeGroupResponse' (Maybe AttributeGroup) Int

newUpdateAttributeGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateAttributeGroupResponse'

-> UpdateAttributeGroupResponse

Create a value of UpdateAttributeGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeGroup:UpdateAttributeGroup', updateAttributeGroupResponse_attributeGroup - The updated information of the attribute group.

$sel:httpStatus:UpdateAttributeGroupResponse', updateAttributeGroupResponse_httpStatus - The response's http status code.

ListAssociatedAttributeGroups (Paginated)

data ListAssociatedAttributeGroups

See: newListAssociatedAttributeGroups smart constructor.

Constructors

ListAssociatedAttributeGroups' (Maybe Text) (Maybe Natural) Text

newListAssociatedAttributeGroups

Arguments

:: Text
$sel:application:ListAssociatedAttributeGroups'

-> ListAssociatedAttributeGroups

Create a value of ListAssociatedAttributeGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociatedAttributeGroups', listAssociatedAttributeGroups_nextToken - The token to use to get the next page of results after a previous API call.

$sel:maxResults:ListAssociatedAttributeGroups', listAssociatedAttributeGroups_maxResults - The upper bound of the number of results to return (cannot exceed 25). If this parameter is omitted, it defaults to 25. This value is optional.

$sel:application:ListAssociatedAttributeGroups', listAssociatedAttributeGroups_application - The name or ID of the application.

data ListAssociatedAttributeGroupsResponse

See: newListAssociatedAttributeGroupsResponse smart constructor.

Constructors

ListAssociatedAttributeGroupsResponse' (Maybe [Text]) (Maybe Text) Int

newListAssociatedAttributeGroupsResponse

Arguments

:: Int
$sel:httpStatus:ListAssociatedAttributeGroupsResponse'

-> ListAssociatedAttributeGroupsResponse

Create a value of ListAssociatedAttributeGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeGroups:ListAssociatedAttributeGroupsResponse', listAssociatedAttributeGroupsResponse_attributeGroups - A list of attribute group IDs.

$sel:nextToken:ListAssociatedAttributeGroups', listAssociatedAttributeGroupsResponse_nextToken - The token to use to get the next page of results after a previous API call.

$sel:httpStatus:ListAssociatedAttributeGroupsResponse', listAssociatedAttributeGroupsResponse_httpStatus - The response's http status code.

GetAttributeGroup

data GetAttributeGroup

See: newGetAttributeGroup smart constructor.

Constructors

GetAttributeGroup' Text

newGetAttributeGroup

Arguments

:: Text
$sel:attributeGroup:GetAttributeGroup'

-> GetAttributeGroup

Create a value of GetAttributeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeGroup:GetAttributeGroup', getAttributeGroup_attributeGroup - The name or ID of the attribute group that holds the attributes to describe the application.

data GetAttributeGroupResponse

See: newGetAttributeGroupResponse smart constructor.

Constructors

GetAttributeGroupResponse' (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe (HashMap Text Text)) Int

newGetAttributeGroupResponse

Arguments

:: Int
$sel:httpStatus:GetAttributeGroupResponse'

-> GetAttributeGroupResponse

Create a value of GetAttributeGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:GetAttributeGroupResponse', getAttributeGroupResponse_creationTime - The ISO-8601 formatted timestamp of the moment the attribute group was created.

$sel:arn:GetAttributeGroupResponse', getAttributeGroupResponse_arn - The Amazon resource name (ARN) that specifies the attribute group across services.

$sel:name:GetAttributeGroupResponse', getAttributeGroupResponse_name - The name of the attribute group.

$sel:attributes:GetAttributeGroupResponse', getAttributeGroupResponse_attributes - A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

$sel:id:GetAttributeGroupResponse', getAttributeGroupResponse_id - The identifier of the attribute group.

$sel:lastUpdateTime:GetAttributeGroupResponse', getAttributeGroupResponse_lastUpdateTime - The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

$sel:description:GetAttributeGroupResponse', getAttributeGroupResponse_description - The description of the attribute group that the user provides.

$sel:tags:GetAttributeGroupResponse', getAttributeGroupResponse_tags - Key-value pairs associated with the attribute group.

$sel:httpStatus:GetAttributeGroupResponse', getAttributeGroupResponse_httpStatus - The response's http status code.

DisassociateResource

data DisassociateResource

See: newDisassociateResource smart constructor.

Constructors

DisassociateResource' Text ResourceType Text

newDisassociateResource

Arguments

:: Text
$sel:application:DisassociateResource'

-> ResourceType
$sel:resourceType:DisassociateResource'

-> Text
$sel:resource:DisassociateResource'

-> DisassociateResource

Create a value of DisassociateResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:DisassociateResource', disassociateResource_application - The name or ID of the application.

$sel:resourceType:DisassociateResource', disassociateResource_resourceType - The type of the resource that is being disassociated.

$sel:resource:DisassociateResource', disassociateResource_resource - The name or ID of the resource.

data DisassociateResourceResponse

See: newDisassociateResourceResponse smart constructor.

Constructors

DisassociateResourceResponse' (Maybe Text) (Maybe Text) Int

newDisassociateResourceResponse

Arguments

:: Int
$sel:httpStatus:DisassociateResourceResponse'

-> DisassociateResourceResponse

Create a value of DisassociateResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationArn:DisassociateResourceResponse', disassociateResourceResponse_applicationArn - The Amazon resource name (ARN) that specifies the application.

$sel:resourceArn:DisassociateResourceResponse', disassociateResourceResponse_resourceArn - The Amazon resource name (ARN) that specifies the resource.

$sel:httpStatus:DisassociateResourceResponse', disassociateResourceResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The Amazon resource name (ARN) that specifies the resource.

$sel:tags:TagResource', tagResource_tags - The new or modified tags for the resource.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

ListAssociatedResources (Paginated)

data ListAssociatedResources

See: newListAssociatedResources smart constructor.

Constructors

ListAssociatedResources' (Maybe Text) (Maybe Natural) Text

newListAssociatedResources

Arguments

:: Text
$sel:application:ListAssociatedResources'

-> ListAssociatedResources

Create a value of ListAssociatedResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssociatedResources', listAssociatedResources_nextToken - The token to use to get the next page of results after a previous API call.

$sel:maxResults:ListAssociatedResources', listAssociatedResources_maxResults - The upper bound of the number of results to return (cannot exceed 25). If this parameter is omitted, it defaults to 25. This value is optional.

$sel:application:ListAssociatedResources', listAssociatedResources_application - The name or ID of the application.

data ListAssociatedResourcesResponse

See: newListAssociatedResourcesResponse smart constructor.

Constructors

ListAssociatedResourcesResponse' (Maybe [ResourceInfo]) (Maybe Text) Int

newListAssociatedResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListAssociatedResourcesResponse'

-> ListAssociatedResourcesResponse

Create a value of ListAssociatedResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resources:ListAssociatedResourcesResponse', listAssociatedResourcesResponse_resources - Information about the resources.

$sel:nextToken:ListAssociatedResources', listAssociatedResourcesResponse_nextToken - The token to use to get the next page of results after a previous API call.

$sel:httpStatus:ListAssociatedResourcesResponse', listAssociatedResourcesResponse_httpStatus - The response's http status code.

ListApplications (Paginated)

data ListApplications

See: newListApplications smart constructor.

Constructors

ListApplications' (Maybe Text) (Maybe Natural)

newListApplications :: ListApplications

Create a value of ListApplications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListApplications', listApplications_nextToken - The token to use to get the next page of results after a previous API call.

$sel:maxResults:ListApplications', listApplications_maxResults - The upper bound of the number of results to return (cannot exceed 25). If this parameter is omitted, it defaults to 25. This value is optional.

data ListApplicationsResponse

See: newListApplicationsResponse smart constructor.

Constructors

ListApplicationsResponse' (Maybe Text) (Maybe [ApplicationSummary]) Int

newListApplicationsResponse

Arguments

:: Int
$sel:httpStatus:ListApplicationsResponse'

-> ListApplicationsResponse

Create a value of ListApplicationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListApplications', listApplicationsResponse_nextToken - The token to use to get the next page of results after a previous API call.

$sel:applications:ListApplicationsResponse', listApplicationsResponse_applications - This list of applications.

$sel:httpStatus:ListApplicationsResponse', listApplicationsResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The Amazon resource name (ARN) that specifies the resource.

$sel:tagKeys:UntagResource', untagResource_tagKeys - A list of the tag keys to remove from the specified resource.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

Types

ResourceGroupState

newtype ResourceGroupState

Constructors

ResourceGroupState'

Fields

fromResourceGroupState :: Text

Bundled Patterns

pattern ResourceGroupState_CREATE_COMPLETE :: ResourceGroupState

pattern ResourceGroupState_CREATE_FAILED :: ResourceGroupState

pattern ResourceGroupState_CREATING :: ResourceGroupState

pattern ResourceGroupState_UPDATE_COMPLETE :: ResourceGroupState

pattern ResourceGroupState_UPDATE_FAILED :: ResourceGroupState

pattern ResourceGroupState_UPDATING :: ResourceGroupState

ResourceType

newtype ResourceType

Constructors

ResourceType'

Fields

fromResourceType :: Text

Bundled Patterns

pattern ResourceType_CFN_STACK :: ResourceType

SyncAction

newtype SyncAction

Constructors

SyncAction'

Fields

fromSyncAction :: Text

Bundled Patterns

pattern SyncAction_NO_ACTION :: SyncAction

pattern SyncAction_START_SYNC :: SyncAction

Application

data Application

Represents a Amazon Web Services Service Catalog AppRegistry application that is the top-level node in a hierarchy of related cloud resource abstractions.

See: newApplication smart constructor.

Constructors

Application' (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe (HashMap Text Text))

newApplication :: Application

Create a value of Application with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Application', application_creationTime - The ISO-8601 formatted timestamp of the moment when the application was created.

$sel:arn:Application', application_arn - The Amazon resource name (ARN) that specifies the application across services.

$sel:name:Application', application_name - The name of the application. The name must be unique in the region in which you are creating the application.

$sel:id:Application', application_id - The identifier of the application.

$sel:lastUpdateTime:Application', application_lastUpdateTime - The ISO-8601 formatted timestamp of the moment when the application was last updated.

$sel:description:Application', application_description - The description of the application.

$sel:tags:Application', application_tags - Key-value pairs you can use to associate with the application.

ApplicationSummary

data ApplicationSummary

Summary of a Amazon Web Services Service Catalog AppRegistry application.

See: newApplicationSummary smart constructor.

Constructors

ApplicationSummary' (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe Text)

newApplicationSummary :: ApplicationSummary

Create a value of ApplicationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ApplicationSummary', applicationSummary_creationTime - The ISO-8601 formatted timestamp of the moment when the application was created.

$sel:arn:ApplicationSummary', applicationSummary_arn - The Amazon resource name (ARN) that specifies the application across services.

$sel:name:ApplicationSummary', applicationSummary_name - The name of the application. The name must be unique in the region in which you are creating the application.

$sel:id:ApplicationSummary', applicationSummary_id - The identifier of the application.

$sel:lastUpdateTime:ApplicationSummary', applicationSummary_lastUpdateTime - The ISO-8601 formatted timestamp of the moment when the application was last updated.

$sel:description:ApplicationSummary', applicationSummary_description - The description of the application.

AttributeGroup

data AttributeGroup

Represents a Amazon Web Services Service Catalog AppRegistry attribute group that is rich metadata which describes an application and its components.

See: newAttributeGroup smart constructor.

Constructors

AttributeGroup' (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe (HashMap Text Text))

newAttributeGroup :: AttributeGroup

Create a value of AttributeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:AttributeGroup', attributeGroup_creationTime - The ISO-8601 formatted timestamp of the moment the attribute group was created.

$sel:arn:AttributeGroup', attributeGroup_arn - The Amazon resource name (ARN) that specifies the attribute group across services.

$sel:name:AttributeGroup', attributeGroup_name - The name of the attribute group.

$sel:id:AttributeGroup', attributeGroup_id - The globally unique attribute group identifier of the attribute group.

$sel:lastUpdateTime:AttributeGroup', attributeGroup_lastUpdateTime - The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

$sel:description:AttributeGroup', attributeGroup_description - The description of the attribute group that the user provides.

$sel:tags:AttributeGroup', attributeGroup_tags - Key-value pairs you can use to associate with the attribute group.

AttributeGroupSummary

data AttributeGroupSummary

Summary of a Amazon Web Services Service Catalog AppRegistry attribute group.

See: newAttributeGroupSummary smart constructor.

Constructors

AttributeGroupSummary' (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe POSIX) (Maybe Text)

newAttributeGroupSummary :: AttributeGroupSummary

Create a value of AttributeGroupSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:AttributeGroupSummary', attributeGroupSummary_creationTime - The ISO-8601 formatted timestamp of the moment the attribute group was created.

$sel:arn:AttributeGroupSummary', attributeGroupSummary_arn - The Amazon resource name (ARN) that specifies the attribute group across services.

$sel:name:AttributeGroupSummary', attributeGroupSummary_name - The name of the attribute group.

$sel:id:AttributeGroupSummary', attributeGroupSummary_id - The globally unique attribute group identifier of the attribute group.

$sel:lastUpdateTime:AttributeGroupSummary', attributeGroupSummary_lastUpdateTime - The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

$sel:description:AttributeGroupSummary', attributeGroupSummary_description - The description of the attribute group that the user provides.

Integrations

data Integrations

The information about the service integration.

See: newIntegrations smart constructor.

Constructors

Integrations' (Maybe ResourceGroup)

newIntegrations :: Integrations

Create a value of Integrations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceGroup:Integrations', integrations_resourceGroup - The information about the resource group integration.

Resource

data Resource

The information about the resource.

See: newResource smart constructor.

Constructors

Resource' (Maybe Text) (Maybe ResourceIntegrations) (Maybe POSIX) (Maybe Text)

newResource :: Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Resource', resource_arn - The Amazon resource name (ARN) of the resource.

$sel:integrations:Resource', resource_integrations - The service integration information about the resource.

$sel:associationTime:Resource', resource_associationTime - The time the resource was associated with the application.

$sel:name:Resource', resource_name - The name of the resource.

ResourceGroup

data ResourceGroup

The information about the resource group integration.

See: newResourceGroup smart constructor.

Constructors

ResourceGroup' (Maybe ResourceGroupState) (Maybe Text) (Maybe Text)

newResourceGroup :: ResourceGroup

Create a value of ResourceGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ResourceGroup', resourceGroup_state - The state of the propagation process for the resource group. The states includes:

CREATING if the resource group is in the process of being created.

CREATE_COMPLETE if the resource group was created successfully.

CREATE_FAILED if the resource group failed to be created.

UPDATING if the resource group is in the process of being updated.

UPDATE_COMPLETE if the resource group updated successfully.

UPDATE_FAILED if the resource group could not update successfully.

$sel:arn:ResourceGroup', resourceGroup_arn - The Amazon resource name (ARN) of the resource group.

$sel:errorMessage:ResourceGroup', resourceGroup_errorMessage - The error message that generates when the propagation process for the resource group fails.

ResourceInfo

data ResourceInfo

The information about the resource.

See: newResourceInfo smart constructor.

Constructors

ResourceInfo' (Maybe Text) (Maybe Text)

newResourceInfo :: ResourceInfo

Create a value of ResourceInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ResourceInfo', resourceInfo_arn - The Amazon resource name (ARN) that specifies the resource across services.

$sel:name:ResourceInfo', resourceInfo_name - The name of the resource.

ResourceIntegrations

data ResourceIntegrations

The service integration information about the resource.

See: newResourceIntegrations smart constructor.

Constructors

ResourceIntegrations' (Maybe ResourceGroup)

newResourceIntegrations :: ResourceIntegrations

Create a value of ResourceIntegrations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceGroup:ResourceIntegrations', resourceIntegrations_resourceGroup - The information about the integration of Resource Groups.