Amazonka.ServiceCatalogAppRegistry.CreateApplication

Description

Creates a new application that is the top-level node in a hierarchy of related cloud resource abstractions.

Creating a Request

data CreateApplication

See: newCreateApplication smart constructor.

Constructors

CreateApplication'

Fields

description :: Maybe Text

The description of the application.

tags :: Maybe (HashMap Text Text)

Key-value pairs you can use to associate with the application.

name :: Text

The name of the application. The name must be unique in the region in which you are creating the application.

clientToken :: Text

A unique identifier that you provide to ensure idempotency. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails.

newCreateApplication

Arguments

:: Text
$sel:name:CreateApplication'

-> Text
$sel:clientToken:CreateApplication'

-> CreateApplication

Create a value of CreateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateApplication', createApplication_description - The description of the application.

$sel:tags:CreateApplication', createApplication_tags - Key-value pairs you can use to associate with the application.

$sel:name:CreateApplication', createApplication_name - The name of the application. The name must be unique in the region in which you are creating the application.

$sel:clientToken:CreateApplication', createApplication_clientToken - A unique identifier that you provide to ensure idempotency. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails.

Request Lenses

createApplication_description :: Lens' CreateApplication (Maybe Text)

The description of the application.

createApplication_tags :: Lens' CreateApplication (Maybe (HashMap Text Text))

Key-value pairs you can use to associate with the application.

createApplication_name :: Lens' CreateApplication Text

The name of the application. The name must be unique in the region in which you are creating the application.

createApplication_clientToken :: Lens' CreateApplication Text

A unique identifier that you provide to ensure idempotency. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails.

Destructuring the Response

data CreateApplicationResponse

See: newCreateApplicationResponse smart constructor.

Constructors

CreateApplicationResponse'

Fields

application :: Maybe Application

Information about the application.

httpStatus :: Int

The response's http status code.

newCreateApplicationResponse

Arguments

:: Int
$sel:httpStatus:CreateApplicationResponse'

-> CreateApplicationResponse

Create a value of CreateApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:CreateApplicationResponse', createApplicationResponse_application - Information about the application.

$sel:httpStatus:CreateApplicationResponse', createApplicationResponse_httpStatus - The response's http status code.

Response Lenses

createApplicationResponse_application :: Lens' CreateApplicationResponse (Maybe Application)

Information about the application.

createApplicationResponse_httpStatus :: Lens' CreateApplicationResponse Int

The response's http status code.