Amazonka.ServiceCatalogAppRegistry.CreateAttributeGroup

Description

Creates a new attribute group as a container for user-defined attributes. This feature enables users to have full control over their cloud application's metadata in a rich machine-readable format to facilitate integration with automated workflows and third-party tools.

Creating a Request

data CreateAttributeGroup

See: newCreateAttributeGroup smart constructor.

Constructors

CreateAttributeGroup'

Fields

description :: Maybe Text

The description of the attribute group that the user provides.

tags :: Maybe (HashMap Text Text)

Key-value pairs you can use to associate with the attribute group.

name :: Text

The name of the attribute group.

attributes :: Text

A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

clientToken :: Text

A unique identifier that you provide to ensure idempotency. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails.

newCreateAttributeGroup

Arguments

:: Text
$sel:name:CreateAttributeGroup'

-> Text
$sel:attributes:CreateAttributeGroup'

-> Text
$sel:clientToken:CreateAttributeGroup'

-> CreateAttributeGroup

Create a value of CreateAttributeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateAttributeGroup', createAttributeGroup_description - The description of the attribute group that the user provides.

$sel:tags:CreateAttributeGroup', createAttributeGroup_tags - Key-value pairs you can use to associate with the attribute group.

$sel:name:CreateAttributeGroup', createAttributeGroup_name - The name of the attribute group.

$sel:attributes:CreateAttributeGroup', createAttributeGroup_attributes - A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

$sel:clientToken:CreateAttributeGroup', createAttributeGroup_clientToken - A unique identifier that you provide to ensure idempotency. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails.

Request Lenses

createAttributeGroup_description :: Lens' CreateAttributeGroup (Maybe Text)

The description of the attribute group that the user provides.

createAttributeGroup_tags :: Lens' CreateAttributeGroup (Maybe (HashMap Text Text))

Key-value pairs you can use to associate with the attribute group.

createAttributeGroup_name :: Lens' CreateAttributeGroup Text

The name of the attribute group.

createAttributeGroup_attributes :: Lens' CreateAttributeGroup Text

A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

createAttributeGroup_clientToken :: Lens' CreateAttributeGroup Text

A unique identifier that you provide to ensure idempotency. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails.

Destructuring the Response

data CreateAttributeGroupResponse

See: newCreateAttributeGroupResponse smart constructor.

Constructors

CreateAttributeGroupResponse'

Fields

attributeGroup :: Maybe AttributeGroup

Information about the attribute group.

httpStatus :: Int

The response's http status code.

newCreateAttributeGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateAttributeGroupResponse'

-> CreateAttributeGroupResponse

Create a value of CreateAttributeGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeGroup:CreateAttributeGroupResponse', createAttributeGroupResponse_attributeGroup - Information about the attribute group.

$sel:httpStatus:CreateAttributeGroupResponse', createAttributeGroupResponse_httpStatus - The response's http status code.

Response Lenses

createAttributeGroupResponse_attributeGroup :: Lens' CreateAttributeGroupResponse (Maybe AttributeGroup)

Information about the attribute group.

createAttributeGroupResponse_httpStatus :: Lens' CreateAttributeGroupResponse Int

The response's http status code.