Amazonka.ServiceCatalogAppRegistry.Lens

Description

Operations

AssociateAttributeGroup

associateAttributeGroup_application :: Lens' AssociateAttributeGroup Text

The name or ID of the application.

associateAttributeGroup_attributeGroup :: Lens' AssociateAttributeGroup Text

The name or ID of the attribute group that holds the attributes to describe the application.

associateAttributeGroupResponse_applicationArn :: Lens' AssociateAttributeGroupResponse (Maybe Text)

The Amazon resource name (ARN) of the application that was augmented with attributes.

associateAttributeGroupResponse_attributeGroupArn :: Lens' AssociateAttributeGroupResponse (Maybe Text)

The Amazon resource name (ARN) of the attribute group that contains the application's new attributes.

associateAttributeGroupResponse_httpStatus :: Lens' AssociateAttributeGroupResponse Int

The response's http status code.

ListAttributeGroups

listAttributeGroups_nextToken :: Lens' ListAttributeGroups (Maybe Text)

The token to use to get the next page of results after a previous API call.

listAttributeGroups_maxResults :: Lens' ListAttributeGroups (Maybe Natural)

The upper bound of the number of results to return (cannot exceed 25). If this parameter is omitted, it defaults to 25. This value is optional.

listAttributeGroupsResponse_attributeGroups :: Lens' ListAttributeGroupsResponse (Maybe [AttributeGroupSummary])

This list of attribute groups.

listAttributeGroupsResponse_nextToken :: Lens' ListAttributeGroupsResponse (Maybe Text)

The token to use to get the next page of results after a previous API call.

listAttributeGroupsResponse_httpStatus :: Lens' ListAttributeGroupsResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon resource name (ARN) that specifies the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags on the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

SyncResource

syncResource_resourceType :: Lens' SyncResource ResourceType

The type of resource of which the application will be associated.

syncResource_resource :: Lens' SyncResource Text

An entity you can work with and specify with a name or ID. Examples include an Amazon EC2 instance, an Amazon Web Services CloudFormation stack, or an Amazon S3 bucket.

syncResourceResponse_applicationArn :: Lens' SyncResourceResponse (Maybe Text)

The Amazon resource name (ARN) that specifies the application.

syncResourceResponse_actionTaken :: Lens' SyncResourceResponse (Maybe SyncAction)

The results of the output if an application is associated with an ARN value, which could be syncStarted or None.

syncResourceResponse_resourceArn :: Lens' SyncResourceResponse (Maybe Text)

The Amazon resource name (ARN) that specifies the resource.

syncResourceResponse_httpStatus :: Lens' SyncResourceResponse Int

The response's http status code.

DeleteApplication

deleteApplication_application :: Lens' DeleteApplication Text

The name or ID of the application.

deleteApplicationResponse_application :: Lens' DeleteApplicationResponse (Maybe ApplicationSummary)

Information about the deleted application.

deleteApplicationResponse_httpStatus :: Lens' DeleteApplicationResponse Int

The response's http status code.

UpdateApplication

updateApplication_name :: Lens' UpdateApplication (Maybe Text)

The new name of the application. The name must be unique in the region in which you are updating the application.

updateApplication_description :: Lens' UpdateApplication (Maybe Text)

The new description of the application.

updateApplication_application :: Lens' UpdateApplication Text

The name or ID of the application that will be updated.

updateApplicationResponse_application :: Lens' UpdateApplicationResponse (Maybe Application)

The updated information of the application.

updateApplicationResponse_httpStatus :: Lens' UpdateApplicationResponse Int

The response's http status code.

AssociateResource

associateResource_application :: Lens' AssociateResource Text

The name or ID of the application.

associateResource_resourceType :: Lens' AssociateResource ResourceType

The type of resource of which the application will be associated.

associateResource_resource :: Lens' AssociateResource Text

The name or ID of the resource of which the application will be associated.

associateResourceResponse_applicationArn :: Lens' AssociateResourceResponse (Maybe Text)

The Amazon resource name (ARN) of the application that was augmented with attributes.

associateResourceResponse_resourceArn :: Lens' AssociateResourceResponse (Maybe Text)

The Amazon resource name (ARN) that specifies the resource.

associateResourceResponse_httpStatus :: Lens' AssociateResourceResponse Int

The response's http status code.

CreateApplication

createApplication_description :: Lens' CreateApplication (Maybe Text)

The description of the application.

createApplication_tags :: Lens' CreateApplication (Maybe (HashMap Text Text))

Key-value pairs you can use to associate with the application.

createApplication_name :: Lens' CreateApplication Text

The name of the application. The name must be unique in the region in which you are creating the application.

createApplication_clientToken :: Lens' CreateApplication Text

A unique identifier that you provide to ensure idempotency. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails.

createApplicationResponse_application :: Lens' CreateApplicationResponse (Maybe Application)

Information about the application.

createApplicationResponse_httpStatus :: Lens' CreateApplicationResponse Int

The response's http status code.

DisassociateAttributeGroup

disassociateAttributeGroup_application :: Lens' DisassociateAttributeGroup Text

The name or ID of the application.

disassociateAttributeGroup_attributeGroup :: Lens' DisassociateAttributeGroup Text

The name or ID of the attribute group that holds the attributes to describe the application.

disassociateAttributeGroupResponse_applicationArn :: Lens' DisassociateAttributeGroupResponse (Maybe Text)

The Amazon resource name (ARN) that specifies the application.

disassociateAttributeGroupResponse_attributeGroupArn :: Lens' DisassociateAttributeGroupResponse (Maybe Text)

The Amazon resource name (ARN) that specifies the attribute group.

disassociateAttributeGroupResponse_httpStatus :: Lens' DisassociateAttributeGroupResponse Int

The response's http status code.

GetApplication

getApplication_application :: Lens' GetApplication Text

The name or ID of the application.

getApplicationResponse_creationTime :: Lens' GetApplicationResponse (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment when the application was created.

getApplicationResponse_arn :: Lens' GetApplicationResponse (Maybe Text)

The Amazon resource name (ARN) that specifies the application across services.

getApplicationResponse_integrations :: Lens' GetApplicationResponse (Maybe Integrations)

The information about the integration of the application with other services, such as Resource Groups.

getApplicationResponse_associatedResourceCount :: Lens' GetApplicationResponse (Maybe Natural)

The number of top-level resources that were registered as part of this application.

getApplicationResponse_name :: Lens' GetApplicationResponse (Maybe Text)

The name of the application. The name must be unique in the region in which you are creating the application.

getApplicationResponse_id :: Lens' GetApplicationResponse (Maybe Text)

The identifier of the application.

getApplicationResponse_lastUpdateTime :: Lens' GetApplicationResponse (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment when the application was last updated.

getApplicationResponse_description :: Lens' GetApplicationResponse (Maybe Text)

The description of the application.

getApplicationResponse_tags :: Lens' GetApplicationResponse (Maybe (HashMap Text Text))

Key-value pairs associated with the application.

getApplicationResponse_httpStatus :: Lens' GetApplicationResponse Int

The response's http status code.

GetAssociatedResource

getAssociatedResource_application :: Lens' GetAssociatedResource Text

The name or ID of the application.

getAssociatedResource_resourceType :: Lens' GetAssociatedResource ResourceType

The type of resource associated with the application.

getAssociatedResource_resource :: Lens' GetAssociatedResource Text

The name or ID of the resource associated with the application.

getAssociatedResourceResponse_resource :: Lens' GetAssociatedResourceResponse (Maybe Resource)

The resource associated with the application.

getAssociatedResourceResponse_httpStatus :: Lens' GetAssociatedResourceResponse Int

The response's http status code.

CreateAttributeGroup

createAttributeGroup_description :: Lens' CreateAttributeGroup (Maybe Text)

The description of the attribute group that the user provides.

createAttributeGroup_tags :: Lens' CreateAttributeGroup (Maybe (HashMap Text Text))

Key-value pairs you can use to associate with the attribute group.

createAttributeGroup_name :: Lens' CreateAttributeGroup Text

The name of the attribute group.

createAttributeGroup_attributes :: Lens' CreateAttributeGroup Text

A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

createAttributeGroup_clientToken :: Lens' CreateAttributeGroup Text

A unique identifier that you provide to ensure idempotency. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails.

createAttributeGroupResponse_attributeGroup :: Lens' CreateAttributeGroupResponse (Maybe AttributeGroup)

Information about the attribute group.

createAttributeGroupResponse_httpStatus :: Lens' CreateAttributeGroupResponse Int

The response's http status code.

DeleteAttributeGroup

deleteAttributeGroup_attributeGroup :: Lens' DeleteAttributeGroup Text

The name or ID of the attribute group that holds the attributes to describe the application.

deleteAttributeGroupResponse_attributeGroup :: Lens' DeleteAttributeGroupResponse (Maybe AttributeGroupSummary)

Information about the deleted attribute group.

deleteAttributeGroupResponse_httpStatus :: Lens' DeleteAttributeGroupResponse Int

The response's http status code.

UpdateAttributeGroup

updateAttributeGroup_name :: Lens' UpdateAttributeGroup (Maybe Text)

The new name of the attribute group. The name must be unique in the region in which you are updating the attribute group.

updateAttributeGroup_attributes :: Lens' UpdateAttributeGroup (Maybe Text)

A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

updateAttributeGroup_description :: Lens' UpdateAttributeGroup (Maybe Text)

The description of the attribute group that the user provides.

updateAttributeGroup_attributeGroup :: Lens' UpdateAttributeGroup Text

The name or ID of the attribute group that holds the attributes to describe the application.

updateAttributeGroupResponse_attributeGroup :: Lens' UpdateAttributeGroupResponse (Maybe AttributeGroup)

The updated information of the attribute group.

updateAttributeGroupResponse_httpStatus :: Lens' UpdateAttributeGroupResponse Int

The response's http status code.

ListAssociatedAttributeGroups

listAssociatedAttributeGroups_nextToken :: Lens' ListAssociatedAttributeGroups (Maybe Text)

The token to use to get the next page of results after a previous API call.

listAssociatedAttributeGroups_maxResults :: Lens' ListAssociatedAttributeGroups (Maybe Natural)

The upper bound of the number of results to return (cannot exceed 25). If this parameter is omitted, it defaults to 25. This value is optional.

listAssociatedAttributeGroups_application :: Lens' ListAssociatedAttributeGroups Text

The name or ID of the application.

listAssociatedAttributeGroupsResponse_attributeGroups :: Lens' ListAssociatedAttributeGroupsResponse (Maybe [Text])

A list of attribute group IDs.

listAssociatedAttributeGroupsResponse_nextToken :: Lens' ListAssociatedAttributeGroupsResponse (Maybe Text)

The token to use to get the next page of results after a previous API call.

listAssociatedAttributeGroupsResponse_httpStatus :: Lens' ListAssociatedAttributeGroupsResponse Int

The response's http status code.

GetAttributeGroup

getAttributeGroup_attributeGroup :: Lens' GetAttributeGroup Text

The name or ID of the attribute group that holds the attributes to describe the application.

getAttributeGroupResponse_creationTime :: Lens' GetAttributeGroupResponse (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment the attribute group was created.

getAttributeGroupResponse_arn :: Lens' GetAttributeGroupResponse (Maybe Text)

The Amazon resource name (ARN) that specifies the attribute group across services.

getAttributeGroupResponse_name :: Lens' GetAttributeGroupResponse (Maybe Text)

The name of the attribute group.

getAttributeGroupResponse_attributes :: Lens' GetAttributeGroupResponse (Maybe Text)

A JSON string in the form of nested key-value pairs that represent the attributes in the group and describes an application and its components.

getAttributeGroupResponse_id :: Lens' GetAttributeGroupResponse (Maybe Text)

The identifier of the attribute group.

getAttributeGroupResponse_lastUpdateTime :: Lens' GetAttributeGroupResponse (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

getAttributeGroupResponse_description :: Lens' GetAttributeGroupResponse (Maybe Text)

The description of the attribute group that the user provides.

getAttributeGroupResponse_tags :: Lens' GetAttributeGroupResponse (Maybe (HashMap Text Text))

Key-value pairs associated with the attribute group.

getAttributeGroupResponse_httpStatus :: Lens' GetAttributeGroupResponse Int

The response's http status code.

DisassociateResource

disassociateResource_application :: Lens' DisassociateResource Text

The name or ID of the application.

disassociateResource_resourceType :: Lens' DisassociateResource ResourceType

The type of the resource that is being disassociated.

disassociateResource_resource :: Lens' DisassociateResource Text

The name or ID of the resource.

disassociateResourceResponse_applicationArn :: Lens' DisassociateResourceResponse (Maybe Text)

The Amazon resource name (ARN) that specifies the application.

disassociateResourceResponse_resourceArn :: Lens' DisassociateResourceResponse (Maybe Text)

The Amazon resource name (ARN) that specifies the resource.

disassociateResourceResponse_httpStatus :: Lens' DisassociateResourceResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon resource name (ARN) that specifies the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The new or modified tags for the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListAssociatedResources

listAssociatedResources_nextToken :: Lens' ListAssociatedResources (Maybe Text)

The token to use to get the next page of results after a previous API call.

listAssociatedResources_maxResults :: Lens' ListAssociatedResources (Maybe Natural)

The upper bound of the number of results to return (cannot exceed 25). If this parameter is omitted, it defaults to 25. This value is optional.

listAssociatedResources_application :: Lens' ListAssociatedResources Text

The name or ID of the application.

listAssociatedResourcesResponse_resources :: Lens' ListAssociatedResourcesResponse (Maybe [ResourceInfo])

Information about the resources.

listAssociatedResourcesResponse_nextToken :: Lens' ListAssociatedResourcesResponse (Maybe Text)

The token to use to get the next page of results after a previous API call.

listAssociatedResourcesResponse_httpStatus :: Lens' ListAssociatedResourcesResponse Int

The response's http status code.

ListApplications

listApplications_nextToken :: Lens' ListApplications (Maybe Text)

The token to use to get the next page of results after a previous API call.

listApplications_maxResults :: Lens' ListApplications (Maybe Natural)

The upper bound of the number of results to return (cannot exceed 25). If this parameter is omitted, it defaults to 25. This value is optional.

listApplicationsResponse_nextToken :: Lens' ListApplicationsResponse (Maybe Text)

The token to use to get the next page of results after a previous API call.

listApplicationsResponse_applications :: Lens' ListApplicationsResponse (Maybe [ApplicationSummary])

This list of applications.

listApplicationsResponse_httpStatus :: Lens' ListApplicationsResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon resource name (ARN) that specifies the resource.

untagResource_tagKeys :: Lens' UntagResource [Text]

A list of the tag keys to remove from the specified resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

Types

Application

application_creationTime :: Lens' Application (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment when the application was created.

application_arn :: Lens' Application (Maybe Text)

The Amazon resource name (ARN) that specifies the application across services.

application_name :: Lens' Application (Maybe Text)

The name of the application. The name must be unique in the region in which you are creating the application.

application_id :: Lens' Application (Maybe Text)

The identifier of the application.

application_lastUpdateTime :: Lens' Application (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment when the application was last updated.

application_description :: Lens' Application (Maybe Text)

The description of the application.

application_tags :: Lens' Application (Maybe (HashMap Text Text))

Key-value pairs you can use to associate with the application.

ApplicationSummary

applicationSummary_creationTime :: Lens' ApplicationSummary (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment when the application was created.

applicationSummary_arn :: Lens' ApplicationSummary (Maybe Text)

The Amazon resource name (ARN) that specifies the application across services.

applicationSummary_name :: Lens' ApplicationSummary (Maybe Text)

The name of the application. The name must be unique in the region in which you are creating the application.

applicationSummary_id :: Lens' ApplicationSummary (Maybe Text)

The identifier of the application.

applicationSummary_lastUpdateTime :: Lens' ApplicationSummary (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment when the application was last updated.

applicationSummary_description :: Lens' ApplicationSummary (Maybe Text)

The description of the application.

AttributeGroup

attributeGroup_creationTime :: Lens' AttributeGroup (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment the attribute group was created.

attributeGroup_arn :: Lens' AttributeGroup (Maybe Text)

The Amazon resource name (ARN) that specifies the attribute group across services.

attributeGroup_name :: Lens' AttributeGroup (Maybe Text)

The name of the attribute group.

attributeGroup_id :: Lens' AttributeGroup (Maybe Text)

The globally unique attribute group identifier of the attribute group.

attributeGroup_lastUpdateTime :: Lens' AttributeGroup (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

attributeGroup_description :: Lens' AttributeGroup (Maybe Text)

The description of the attribute group that the user provides.

attributeGroup_tags :: Lens' AttributeGroup (Maybe (HashMap Text Text))

Key-value pairs you can use to associate with the attribute group.

AttributeGroupSummary

attributeGroupSummary_creationTime :: Lens' AttributeGroupSummary (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment the attribute group was created.

attributeGroupSummary_arn :: Lens' AttributeGroupSummary (Maybe Text)

The Amazon resource name (ARN) that specifies the attribute group across services.

attributeGroupSummary_name :: Lens' AttributeGroupSummary (Maybe Text)

The name of the attribute group.

attributeGroupSummary_id :: Lens' AttributeGroupSummary (Maybe Text)

The globally unique attribute group identifier of the attribute group.

attributeGroupSummary_lastUpdateTime :: Lens' AttributeGroupSummary (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

attributeGroupSummary_description :: Lens' AttributeGroupSummary (Maybe Text)

The description of the attribute group that the user provides.

Integrations

integrations_resourceGroup :: Lens' Integrations (Maybe ResourceGroup)

The information about the resource group integration.

Resource

resource_arn :: Lens' Resource (Maybe Text)

The Amazon resource name (ARN) of the resource.

resource_integrations :: Lens' Resource (Maybe ResourceIntegrations)

The service integration information about the resource.

resource_associationTime :: Lens' Resource (Maybe UTCTime)

The time the resource was associated with the application.

resource_name :: Lens' Resource (Maybe Text)

The name of the resource.

ResourceGroup

resourceGroup_state :: Lens' ResourceGroup (Maybe ResourceGroupState)

The state of the propagation process for the resource group. The states includes:

CREATING if the resource group is in the process of being created.

CREATE_COMPLETE if the resource group was created successfully.

CREATE_FAILED if the resource group failed to be created.

UPDATING if the resource group is in the process of being updated.

UPDATE_COMPLETE if the resource group updated successfully.

UPDATE_FAILED if the resource group could not update successfully.

resourceGroup_arn :: Lens' ResourceGroup (Maybe Text)

The Amazon resource name (ARN) of the resource group.

resourceGroup_errorMessage :: Lens' ResourceGroup (Maybe Text)

The error message that generates when the propagation process for the resource group fails.

ResourceInfo

resourceInfo_arn :: Lens' ResourceInfo (Maybe Text)

The Amazon resource name (ARN) that specifies the resource across services.

resourceInfo_name :: Lens' ResourceInfo (Maybe Text)

The name of the resource.

ResourceIntegrations

resourceIntegrations_resourceGroup :: Lens' ResourceIntegrations (Maybe ResourceGroup)

The information about the integration of Resource Groups.