Amazonka.ServiceCatalogAppRegistry.SyncResource

Description

Syncs the resource with current AppRegistry records.

Specifically, the resource’s AppRegistry system tags sync with its associated application. We remove the resource's AppRegistry system tags if it does not associate with the application. The caller must have permissions to read and update the resource.

Creating a Request

data SyncResource

See: newSyncResource smart constructor.

Constructors

SyncResource'

Fields

resourceType :: ResourceType

The type of resource of which the application will be associated.

resource :: Text

An entity you can work with and specify with a name or ID. Examples include an Amazon EC2 instance, an Amazon Web Services CloudFormation stack, or an Amazon S3 bucket.

newSyncResource

Arguments

:: ResourceType
$sel:resourceType:SyncResource'

-> Text
$sel:resource:SyncResource'

-> SyncResource

Create a value of SyncResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:SyncResource', syncResource_resourceType - The type of resource of which the application will be associated.

$sel:resource:SyncResource', syncResource_resource - An entity you can work with and specify with a name or ID. Examples include an Amazon EC2 instance, an Amazon Web Services CloudFormation stack, or an Amazon S3 bucket.

Request Lenses

syncResource_resourceType :: Lens' SyncResource ResourceType

The type of resource of which the application will be associated.

syncResource_resource :: Lens' SyncResource Text

An entity you can work with and specify with a name or ID. Examples include an Amazon EC2 instance, an Amazon Web Services CloudFormation stack, or an Amazon S3 bucket.

Destructuring the Response

data SyncResourceResponse

See: newSyncResourceResponse smart constructor.

Constructors

SyncResourceResponse'

Fields

applicationArn :: Maybe Text

The Amazon resource name (ARN) that specifies the application.

actionTaken :: Maybe SyncAction

The results of the output if an application is associated with an ARN value, which could be syncStarted or None.

resourceArn :: Maybe Text

The Amazon resource name (ARN) that specifies the resource.

httpStatus :: Int

The response's http status code.

newSyncResourceResponse

Arguments

:: Int
$sel:httpStatus:SyncResourceResponse'

-> SyncResourceResponse

Create a value of SyncResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationArn:SyncResourceResponse', syncResourceResponse_applicationArn - The Amazon resource name (ARN) that specifies the application.

$sel:actionTaken:SyncResourceResponse', syncResourceResponse_actionTaken - The results of the output if an application is associated with an ARN value, which could be syncStarted or None.

$sel:resourceArn:SyncResourceResponse', syncResourceResponse_resourceArn - The Amazon resource name (ARN) that specifies the resource.

$sel:httpStatus:SyncResourceResponse', syncResourceResponse_httpStatus - The response's http status code.

Response Lenses

syncResourceResponse_applicationArn :: Lens' SyncResourceResponse (Maybe Text)

The Amazon resource name (ARN) that specifies the application.

syncResourceResponse_actionTaken :: Lens' SyncResourceResponse (Maybe SyncAction)

The results of the output if an application is associated with an ARN value, which could be syncStarted or None.

syncResourceResponse_resourceArn :: Lens' SyncResourceResponse (Maybe Text)

The Amazon resource name (ARN) that specifies the resource.

syncResourceResponse_httpStatus :: Lens' SyncResourceResponse Int

The response's http status code.