Amazonka.ServiceCatalogAppRegistry.Types

Description

Service Configuration

defaultService :: Service

API version 2020-06-24 of the Amazon Service Catalog App Registry SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The request has invalid or missing parameters.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

There was a conflict when processing the request (for example, a resource with the given name already exists within the account).

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of resources per account has been reached.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The service is experiencing internal problems.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

ResourceGroupState

newtype ResourceGroupState

Constructors

ResourceGroupState'

Fields

fromResourceGroupState :: Text

Bundled Patterns

pattern ResourceGroupState_CREATE_COMPLETE :: ResourceGroupState

pattern ResourceGroupState_CREATE_FAILED :: ResourceGroupState

pattern ResourceGroupState_CREATING :: ResourceGroupState

pattern ResourceGroupState_UPDATE_COMPLETE :: ResourceGroupState

pattern ResourceGroupState_UPDATE_FAILED :: ResourceGroupState

pattern ResourceGroupState_UPDATING :: ResourceGroupState

ResourceType

newtype ResourceType

Constructors

ResourceType'

Fields

fromResourceType :: Text

Bundled Patterns

pattern ResourceType_CFN_STACK :: ResourceType

SyncAction

newtype SyncAction

Constructors

SyncAction'

Fields

fromSyncAction :: Text

Bundled Patterns

pattern SyncAction_NO_ACTION :: SyncAction

pattern SyncAction_START_SYNC :: SyncAction

Application

data Application

Represents a Amazon Web Services Service Catalog AppRegistry application that is the top-level node in a hierarchy of related cloud resource abstractions.

See: newApplication smart constructor.

Constructors

Application'

Fields

creationTime :: Maybe POSIX

The ISO-8601 formatted timestamp of the moment when the application was created.

arn :: Maybe Text

The Amazon resource name (ARN) that specifies the application across services.

name :: Maybe Text

The name of the application. The name must be unique in the region in which you are creating the application.

id :: Maybe Text

The identifier of the application.

lastUpdateTime :: Maybe POSIX

The ISO-8601 formatted timestamp of the moment when the application was last updated.

description :: Maybe Text

The description of the application.

tags :: Maybe (HashMap Text Text)

Key-value pairs you can use to associate with the application.

newApplication :: Application

Create a value of Application with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Application', application_creationTime - The ISO-8601 formatted timestamp of the moment when the application was created.

$sel:arn:Application', application_arn - The Amazon resource name (ARN) that specifies the application across services.

$sel:name:Application', application_name - The name of the application. The name must be unique in the region in which you are creating the application.

$sel:id:Application', application_id - The identifier of the application.

$sel:lastUpdateTime:Application', application_lastUpdateTime - The ISO-8601 formatted timestamp of the moment when the application was last updated.

$sel:description:Application', application_description - The description of the application.

$sel:tags:Application', application_tags - Key-value pairs you can use to associate with the application.

application_creationTime :: Lens' Application (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment when the application was created.

application_arn :: Lens' Application (Maybe Text)

The Amazon resource name (ARN) that specifies the application across services.

application_name :: Lens' Application (Maybe Text)

The name of the application. The name must be unique in the region in which you are creating the application.

application_id :: Lens' Application (Maybe Text)

The identifier of the application.

application_lastUpdateTime :: Lens' Application (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment when the application was last updated.

application_description :: Lens' Application (Maybe Text)

The description of the application.

application_tags :: Lens' Application (Maybe (HashMap Text Text))

Key-value pairs you can use to associate with the application.

ApplicationSummary

data ApplicationSummary

Summary of a Amazon Web Services Service Catalog AppRegistry application.

See: newApplicationSummary smart constructor.

Constructors

ApplicationSummary'

Fields

creationTime :: Maybe POSIX

The ISO-8601 formatted timestamp of the moment when the application was created.

arn :: Maybe Text

The Amazon resource name (ARN) that specifies the application across services.

name :: Maybe Text

The name of the application. The name must be unique in the region in which you are creating the application.

id :: Maybe Text

The identifier of the application.

lastUpdateTime :: Maybe POSIX

The ISO-8601 formatted timestamp of the moment when the application was last updated.

description :: Maybe Text

The description of the application.

newApplicationSummary :: ApplicationSummary

Create a value of ApplicationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ApplicationSummary', applicationSummary_creationTime - The ISO-8601 formatted timestamp of the moment when the application was created.

$sel:arn:ApplicationSummary', applicationSummary_arn - The Amazon resource name (ARN) that specifies the application across services.

$sel:name:ApplicationSummary', applicationSummary_name - The name of the application. The name must be unique in the region in which you are creating the application.

$sel:id:ApplicationSummary', applicationSummary_id - The identifier of the application.

$sel:lastUpdateTime:ApplicationSummary', applicationSummary_lastUpdateTime - The ISO-8601 formatted timestamp of the moment when the application was last updated.

$sel:description:ApplicationSummary', applicationSummary_description - The description of the application.

applicationSummary_creationTime :: Lens' ApplicationSummary (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment when the application was created.

applicationSummary_arn :: Lens' ApplicationSummary (Maybe Text)

The Amazon resource name (ARN) that specifies the application across services.

applicationSummary_name :: Lens' ApplicationSummary (Maybe Text)

The name of the application. The name must be unique in the region in which you are creating the application.

applicationSummary_id :: Lens' ApplicationSummary (Maybe Text)

The identifier of the application.

applicationSummary_lastUpdateTime :: Lens' ApplicationSummary (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment when the application was last updated.

applicationSummary_description :: Lens' ApplicationSummary (Maybe Text)

The description of the application.

AttributeGroup

data AttributeGroup

Represents a Amazon Web Services Service Catalog AppRegistry attribute group that is rich metadata which describes an application and its components.

See: newAttributeGroup smart constructor.

Constructors

AttributeGroup'

Fields

creationTime :: Maybe POSIX

The ISO-8601 formatted timestamp of the moment the attribute group was created.

arn :: Maybe Text

The Amazon resource name (ARN) that specifies the attribute group across services.

name :: Maybe Text

The name of the attribute group.

id :: Maybe Text

The globally unique attribute group identifier of the attribute group.

lastUpdateTime :: Maybe POSIX

The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

description :: Maybe Text

The description of the attribute group that the user provides.

tags :: Maybe (HashMap Text Text)

Key-value pairs you can use to associate with the attribute group.

newAttributeGroup :: AttributeGroup

Create a value of AttributeGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:AttributeGroup', attributeGroup_creationTime - The ISO-8601 formatted timestamp of the moment the attribute group was created.

$sel:arn:AttributeGroup', attributeGroup_arn - The Amazon resource name (ARN) that specifies the attribute group across services.

$sel:name:AttributeGroup', attributeGroup_name - The name of the attribute group.

$sel:id:AttributeGroup', attributeGroup_id - The globally unique attribute group identifier of the attribute group.

$sel:lastUpdateTime:AttributeGroup', attributeGroup_lastUpdateTime - The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

$sel:description:AttributeGroup', attributeGroup_description - The description of the attribute group that the user provides.

$sel:tags:AttributeGroup', attributeGroup_tags - Key-value pairs you can use to associate with the attribute group.

attributeGroup_creationTime :: Lens' AttributeGroup (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment the attribute group was created.

attributeGroup_arn :: Lens' AttributeGroup (Maybe Text)

The Amazon resource name (ARN) that specifies the attribute group across services.

attributeGroup_name :: Lens' AttributeGroup (Maybe Text)

The name of the attribute group.

attributeGroup_id :: Lens' AttributeGroup (Maybe Text)

The globally unique attribute group identifier of the attribute group.

attributeGroup_lastUpdateTime :: Lens' AttributeGroup (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

attributeGroup_description :: Lens' AttributeGroup (Maybe Text)

The description of the attribute group that the user provides.

attributeGroup_tags :: Lens' AttributeGroup (Maybe (HashMap Text Text))

Key-value pairs you can use to associate with the attribute group.

AttributeGroupSummary

data AttributeGroupSummary

Summary of a Amazon Web Services Service Catalog AppRegistry attribute group.

See: newAttributeGroupSummary smart constructor.

Constructors

AttributeGroupSummary'

Fields

creationTime :: Maybe POSIX

The ISO-8601 formatted timestamp of the moment the attribute group was created.

arn :: Maybe Text

The Amazon resource name (ARN) that specifies the attribute group across services.

name :: Maybe Text

The name of the attribute group.

id :: Maybe Text

The globally unique attribute group identifier of the attribute group.

lastUpdateTime :: Maybe POSIX

The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

description :: Maybe Text

The description of the attribute group that the user provides.

newAttributeGroupSummary :: AttributeGroupSummary

Create a value of AttributeGroupSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:AttributeGroupSummary', attributeGroupSummary_creationTime - The ISO-8601 formatted timestamp of the moment the attribute group was created.

$sel:arn:AttributeGroupSummary', attributeGroupSummary_arn - The Amazon resource name (ARN) that specifies the attribute group across services.

$sel:name:AttributeGroupSummary', attributeGroupSummary_name - The name of the attribute group.

$sel:id:AttributeGroupSummary', attributeGroupSummary_id - The globally unique attribute group identifier of the attribute group.

$sel:lastUpdateTime:AttributeGroupSummary', attributeGroupSummary_lastUpdateTime - The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

$sel:description:AttributeGroupSummary', attributeGroupSummary_description - The description of the attribute group that the user provides.

attributeGroupSummary_creationTime :: Lens' AttributeGroupSummary (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment the attribute group was created.

attributeGroupSummary_arn :: Lens' AttributeGroupSummary (Maybe Text)

The Amazon resource name (ARN) that specifies the attribute group across services.

attributeGroupSummary_name :: Lens' AttributeGroupSummary (Maybe Text)

The name of the attribute group.

attributeGroupSummary_id :: Lens' AttributeGroupSummary (Maybe Text)

The globally unique attribute group identifier of the attribute group.

attributeGroupSummary_lastUpdateTime :: Lens' AttributeGroupSummary (Maybe UTCTime)

The ISO-8601 formatted timestamp of the moment the attribute group was last updated. This time is the same as the creationTime for a newly created attribute group.

attributeGroupSummary_description :: Lens' AttributeGroupSummary (Maybe Text)

The description of the attribute group that the user provides.

Integrations

data Integrations

The information about the service integration.

See: newIntegrations smart constructor.

Constructors

Integrations'

Fields

resourceGroup :: Maybe ResourceGroup

The information about the resource group integration.

newIntegrations :: Integrations

Create a value of Integrations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceGroup:Integrations', integrations_resourceGroup - The information about the resource group integration.

integrations_resourceGroup :: Lens' Integrations (Maybe ResourceGroup)

The information about the resource group integration.

Resource

data Resource

The information about the resource.

See: newResource smart constructor.

Constructors

Resource'

Fields

arn :: Maybe Text

The Amazon resource name (ARN) of the resource.

integrations :: Maybe ResourceIntegrations

The service integration information about the resource.

associationTime :: Maybe POSIX

The time the resource was associated with the application.

name :: Maybe Text

The name of the resource.

newResource :: Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Resource', resource_arn - The Amazon resource name (ARN) of the resource.

$sel:integrations:Resource', resource_integrations - The service integration information about the resource.

$sel:associationTime:Resource', resource_associationTime - The time the resource was associated with the application.

$sel:name:Resource', resource_name - The name of the resource.

resource_arn :: Lens' Resource (Maybe Text)

The Amazon resource name (ARN) of the resource.

resource_integrations :: Lens' Resource (Maybe ResourceIntegrations)

The service integration information about the resource.

resource_associationTime :: Lens' Resource (Maybe UTCTime)

The time the resource was associated with the application.

resource_name :: Lens' Resource (Maybe Text)

The name of the resource.

ResourceGroup

data ResourceGroup

The information about the resource group integration.

See: newResourceGroup smart constructor.

Constructors

ResourceGroup'

Fields

state :: Maybe ResourceGroupState

The state of the propagation process for the resource group. The states includes:

CREATING if the resource group is in the process of being created.

CREATE_COMPLETE if the resource group was created successfully.

CREATE_FAILED if the resource group failed to be created.

UPDATING if the resource group is in the process of being updated.

UPDATE_COMPLETE if the resource group updated successfully.

UPDATE_FAILED if the resource group could not update successfully.

arn :: Maybe Text

The Amazon resource name (ARN) of the resource group.

errorMessage :: Maybe Text

The error message that generates when the propagation process for the resource group fails.

newResourceGroup :: ResourceGroup

Create a value of ResourceGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ResourceGroup', resourceGroup_state - The state of the propagation process for the resource group. The states includes:

CREATING if the resource group is in the process of being created.

CREATE_COMPLETE if the resource group was created successfully.

CREATE_FAILED if the resource group failed to be created.

UPDATING if the resource group is in the process of being updated.

UPDATE_COMPLETE if the resource group updated successfully.

UPDATE_FAILED if the resource group could not update successfully.

$sel:arn:ResourceGroup', resourceGroup_arn - The Amazon resource name (ARN) of the resource group.

$sel:errorMessage:ResourceGroup', resourceGroup_errorMessage - The error message that generates when the propagation process for the resource group fails.

resourceGroup_state :: Lens' ResourceGroup (Maybe ResourceGroupState)

The state of the propagation process for the resource group. The states includes:

CREATING if the resource group is in the process of being created.

CREATE_COMPLETE if the resource group was created successfully.

CREATE_FAILED if the resource group failed to be created.

UPDATING if the resource group is in the process of being updated.

UPDATE_COMPLETE if the resource group updated successfully.

UPDATE_FAILED if the resource group could not update successfully.

resourceGroup_arn :: Lens' ResourceGroup (Maybe Text)

The Amazon resource name (ARN) of the resource group.

resourceGroup_errorMessage :: Lens' ResourceGroup (Maybe Text)

The error message that generates when the propagation process for the resource group fails.

ResourceInfo

data ResourceInfo

The information about the resource.

See: newResourceInfo smart constructor.

Constructors

ResourceInfo'

Fields

arn :: Maybe Text

The Amazon resource name (ARN) that specifies the resource across services.

name :: Maybe Text

The name of the resource.

newResourceInfo :: ResourceInfo

Create a value of ResourceInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ResourceInfo', resourceInfo_arn - The Amazon resource name (ARN) that specifies the resource across services.

$sel:name:ResourceInfo', resourceInfo_name - The name of the resource.

resourceInfo_arn :: Lens' ResourceInfo (Maybe Text)

The Amazon resource name (ARN) that specifies the resource across services.

resourceInfo_name :: Lens' ResourceInfo (Maybe Text)

The name of the resource.

ResourceIntegrations

data ResourceIntegrations

The service integration information about the resource.

See: newResourceIntegrations smart constructor.

Constructors

ResourceIntegrations'

Fields

resourceGroup :: Maybe ResourceGroup

The information about the integration of Resource Groups.

newResourceIntegrations :: ResourceIntegrations

Create a value of ResourceIntegrations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceGroup:ResourceIntegrations', resourceIntegrations_resourceGroup - The information about the integration of Resource Groups.

resourceIntegrations_resourceGroup :: Lens' ResourceIntegrations (Maybe ResourceGroup)

The information about the integration of Resource Groups.