Amazonka.ServiceCatalogAppRegistry.UpdateApplication

Description

Updates an existing application with new attributes.

Creating a Request

data UpdateApplication

See: newUpdateApplication smart constructor.

Constructors

UpdateApplication'

Fields

name :: Maybe Text

The new name of the application. The name must be unique in the region in which you are updating the application.

description :: Maybe Text

The new description of the application.

application :: Text

The name or ID of the application that will be updated.

newUpdateApplication

Arguments

:: Text
$sel:application:UpdateApplication'

-> UpdateApplication

Create a value of UpdateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateApplication', updateApplication_name - The new name of the application. The name must be unique in the region in which you are updating the application.

$sel:description:UpdateApplication', updateApplication_description - The new description of the application.

$sel:application:UpdateApplication', updateApplication_application - The name or ID of the application that will be updated.

Request Lenses

updateApplication_name :: Lens' UpdateApplication (Maybe Text)

The new name of the application. The name must be unique in the region in which you are updating the application.

updateApplication_description :: Lens' UpdateApplication (Maybe Text)

The new description of the application.

updateApplication_application :: Lens' UpdateApplication Text

The name or ID of the application that will be updated.

Destructuring the Response

data UpdateApplicationResponse

See: newUpdateApplicationResponse smart constructor.

Constructors

UpdateApplicationResponse'

Fields

application :: Maybe Application

The updated information of the application.

httpStatus :: Int

The response's http status code.

newUpdateApplicationResponse

Arguments

:: Int
$sel:httpStatus:UpdateApplicationResponse'

-> UpdateApplicationResponse

Create a value of UpdateApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:application:UpdateApplication', updateApplicationResponse_application - The updated information of the application.

$sel:httpStatus:UpdateApplicationResponse', updateApplicationResponse_httpStatus - The response's http status code.

Response Lenses

updateApplicationResponse_application :: Lens' UpdateApplicationResponse (Maybe Application)

The updated information of the application.

updateApplicationResponse_httpStatus :: Lens' UpdateApplicationResponse Int

The response's http status code.