Amazonka.ServiceQuotas

Description

Derived from API version 2019-06-24 of the AWS service descriptions, licensed under Apache 2.0.

With Service Quotas, you can view and manage your quotas easily as your AWS workloads grow. Quotas, also referred to as limits, are the maximum number of resources that you can create in your AWS account. For more information, see the Service Quotas User Guide.

Service Configuration

defaultService :: Service

API version 2019-06-24 of the Amazon Quotas SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ServiceQuotas.

TagPolicyViolationException

_TagPolicyViolationException :: AsError a => Getting (First ServiceError) a ServiceError

The specified tag is a reserved word and cannot be used.

NoAvailableOrganizationException

_NoAvailableOrganizationException :: AsError a => Getting (First ServiceError) a ServiceError

The account making this call is not a member of an organization.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient permission to perform this action.

TemplatesNotAvailableInRegionException

_TemplatesNotAvailableInRegionException :: AsError a => Getting (First ServiceError) a ServiceError

The Service Quotas template is not available in this AWS Region.

DependencyAccessDeniedException

_DependencyAccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You can't perform this action because a dependency does not have access.

ResourceAlreadyExistsException

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource already exists.

AWSServiceAccessNotEnabledException

_AWSServiceAccessNotEnabledException :: AsError a => Getting (First ServiceError) a ServiceError

The action you attempted is not allowed unless Service Access with Service Quotas is enabled in your organization.

InvalidResourceStateException

_InvalidResourceStateException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is in an invalid state.

TooManyTagsException

_TooManyTagsException :: AsError a => Getting (First ServiceError) a ServiceError

You've exceeded the number of tags allowed for a resource. For more information, see Tag restrictions in the Service Quotas User Guide.

TooManyRequestsException

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Due to throttling, the request was denied. Slow down the rate of request calls, or request an increase for this quota.

QuotaExceededException

_QuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded your service quota. To perform the requested action, remove some of the relevant resources, or use Service Quotas to request a service quota increase.

ServiceException

_ServiceException :: AsError a => Getting (First ServiceError) a ServiceError

Something went wrong.

IllegalArgumentException

_IllegalArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

Invalid input was provided.

ServiceQuotaTemplateNotInUseException

_ServiceQuotaTemplateNotInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The quota request template is not associated with your organization.

OrganizationNotInAllFeaturesModeException

_OrganizationNotInAllFeaturesModeException :: AsError a => Getting (First ServiceError) a ServiceError

The organization that your account belongs to is not in All Features mode.

InvalidPaginationTokenException

_InvalidPaginationTokenException :: AsError a => Getting (First ServiceError) a ServiceError

Invalid input was provided.

NoSuchResourceException

_NoSuchResourceException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListServices (Paginated)

data ListServices

See: newListServices smart constructor.

Constructors

ListServices' (Maybe Text) (Maybe Natural)

newListServices :: ListServices

Create a value of ListServices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListServices', listServices_nextToken - The token for the next page of results.

$sel:maxResults:ListServices', listServices_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

data ListServicesResponse

See: newListServicesResponse smart constructor.

Constructors

ListServicesResponse' (Maybe Text) (Maybe [ServiceInfo]) Int

newListServicesResponse

Arguments

:: Int
$sel:httpStatus:ListServicesResponse'

-> ListServicesResponse

Create a value of ListServicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListServices', listServicesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:services:ListServicesResponse', listServicesResponse_services - Information about the services.

$sel:httpStatus:ListServicesResponse', listServicesResponse_httpStatus - The response's http status code.

ListAWSDefaultServiceQuotas (Paginated)

data ListAWSDefaultServiceQuotas

See: newListAWSDefaultServiceQuotas smart constructor.

Constructors

ListAWSDefaultServiceQuotas' (Maybe Text) (Maybe Natural) Text

newListAWSDefaultServiceQuotas

Arguments

:: Text
$sel:serviceCode:ListAWSDefaultServiceQuotas'

-> ListAWSDefaultServiceQuotas

Create a value of ListAWSDefaultServiceQuotas with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAWSDefaultServiceQuotas', listAWSDefaultServiceQuotas_nextToken - The token for the next page of results.

$sel:maxResults:ListAWSDefaultServiceQuotas', listAWSDefaultServiceQuotas_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

$sel:serviceCode:ListAWSDefaultServiceQuotas', listAWSDefaultServiceQuotas_serviceCode - The service identifier.

data ListAWSDefaultServiceQuotasResponse

See: newListAWSDefaultServiceQuotasResponse smart constructor.

Constructors

ListAWSDefaultServiceQuotasResponse' (Maybe Text) (Maybe [ServiceQuota]) Int

newListAWSDefaultServiceQuotasResponse

Arguments

:: Int
$sel:httpStatus:ListAWSDefaultServiceQuotasResponse'

-> ListAWSDefaultServiceQuotasResponse

Create a value of ListAWSDefaultServiceQuotasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAWSDefaultServiceQuotas', listAWSDefaultServiceQuotasResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:quotas:ListAWSDefaultServiceQuotasResponse', listAWSDefaultServiceQuotasResponse_quotas - Information about the quotas.

$sel:httpStatus:ListAWSDefaultServiceQuotasResponse', listAWSDefaultServiceQuotasResponse_httpStatus - The response's http status code.

GetAssociationForServiceQuotaTemplate

data GetAssociationForServiceQuotaTemplate

See: newGetAssociationForServiceQuotaTemplate smart constructor.

Constructors

GetAssociationForServiceQuotaTemplate'

newGetAssociationForServiceQuotaTemplate :: GetAssociationForServiceQuotaTemplate

Create a value of GetAssociationForServiceQuotaTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

data GetAssociationForServiceQuotaTemplateResponse

See: newGetAssociationForServiceQuotaTemplateResponse smart constructor.

Constructors

GetAssociationForServiceQuotaTemplateResponse' (Maybe ServiceQuotaTemplateAssociationStatus) Int

newGetAssociationForServiceQuotaTemplateResponse

Arguments

:: Int
$sel:httpStatus:GetAssociationForServiceQuotaTemplateResponse'

-> GetAssociationForServiceQuotaTemplateResponse

Create a value of GetAssociationForServiceQuotaTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceQuotaTemplateAssociationStatus:GetAssociationForServiceQuotaTemplateResponse', getAssociationForServiceQuotaTemplateResponse_serviceQuotaTemplateAssociationStatus - The association status. If the status is ASSOCIATED, the quota increase requests in the template are automatically applied to new accounts in your organization.

$sel:httpStatus:GetAssociationForServiceQuotaTemplateResponse', getAssociationForServiceQuotaTemplateResponse_httpStatus - The response's http status code.

AssociateServiceQuotaTemplate

data AssociateServiceQuotaTemplate

See: newAssociateServiceQuotaTemplate smart constructor.

Constructors

AssociateServiceQuotaTemplate'

newAssociateServiceQuotaTemplate :: AssociateServiceQuotaTemplate

Create a value of AssociateServiceQuotaTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

data AssociateServiceQuotaTemplateResponse

See: newAssociateServiceQuotaTemplateResponse smart constructor.

Constructors

AssociateServiceQuotaTemplateResponse' Int

newAssociateServiceQuotaTemplateResponse

Arguments

:: Int
$sel:httpStatus:AssociateServiceQuotaTemplateResponse'

-> AssociateServiceQuotaTemplateResponse

Create a value of AssociateServiceQuotaTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateServiceQuotaTemplateResponse', associateServiceQuotaTemplateResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceARN:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARN:ListTagsForResource', listTagsForResource_resourceARN - The Amazon Resource Name (ARN) for the applied quota for which you want to list tags. You can get this information by using the Service Quotas console, or by listing the quotas using the list-service-quotas AWS CLI command or the ListServiceQuotas AWS API operation.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe [Tag]) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - A complex data type that contains zero or more tag elements.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

GetServiceQuota

data GetServiceQuota

See: newGetServiceQuota smart constructor.

Constructors

GetServiceQuota' Text Text

newGetServiceQuota

Arguments

:: Text
$sel:serviceCode:GetServiceQuota'

-> Text
$sel:quotaCode:GetServiceQuota'

-> GetServiceQuota

Create a value of GetServiceQuota with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceCode:GetServiceQuota', getServiceQuota_serviceCode - The service identifier.

$sel:quotaCode:GetServiceQuota', getServiceQuota_quotaCode - The quota identifier.

data GetServiceQuotaResponse

See: newGetServiceQuotaResponse smart constructor.

Constructors

GetServiceQuotaResponse' (Maybe ServiceQuota) Int

newGetServiceQuotaResponse

Arguments

:: Int
$sel:httpStatus:GetServiceQuotaResponse'

-> GetServiceQuotaResponse

Create a value of GetServiceQuotaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quota:GetServiceQuotaResponse', getServiceQuotaResponse_quota - Information about the quota.

$sel:httpStatus:GetServiceQuotaResponse', getServiceQuotaResponse_httpStatus - The response's http status code.

PutServiceQuotaIncreaseRequestIntoTemplate

data PutServiceQuotaIncreaseRequestIntoTemplate

See: newPutServiceQuotaIncreaseRequestIntoTemplate smart constructor.

Constructors

PutServiceQuotaIncreaseRequestIntoTemplate' Text Text Text Double

newPutServiceQuotaIncreaseRequestIntoTemplate

Arguments

:: Text
$sel:quotaCode:PutServiceQuotaIncreaseRequestIntoTemplate'

-> Text
$sel:serviceCode:PutServiceQuotaIncreaseRequestIntoTemplate'

-> Text
$sel:awsRegion:PutServiceQuotaIncreaseRequestIntoTemplate'

-> Double
$sel:desiredValue:PutServiceQuotaIncreaseRequestIntoTemplate'

-> PutServiceQuotaIncreaseRequestIntoTemplate

Create a value of PutServiceQuotaIncreaseRequestIntoTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quotaCode:PutServiceQuotaIncreaseRequestIntoTemplate', putServiceQuotaIncreaseRequestIntoTemplate_quotaCode - The quota identifier.

$sel:serviceCode:PutServiceQuotaIncreaseRequestIntoTemplate', putServiceQuotaIncreaseRequestIntoTemplate_serviceCode - The service identifier.

$sel:awsRegion:PutServiceQuotaIncreaseRequestIntoTemplate', putServiceQuotaIncreaseRequestIntoTemplate_awsRegion - The AWS Region.

$sel:desiredValue:PutServiceQuotaIncreaseRequestIntoTemplate', putServiceQuotaIncreaseRequestIntoTemplate_desiredValue - The new, increased value for the quota.

data PutServiceQuotaIncreaseRequestIntoTemplateResponse

See: newPutServiceQuotaIncreaseRequestIntoTemplateResponse smart constructor.

Constructors

PutServiceQuotaIncreaseRequestIntoTemplateResponse' (Maybe ServiceQuotaIncreaseRequestInTemplate) Int

newPutServiceQuotaIncreaseRequestIntoTemplateResponse

Arguments

:: Int
$sel:httpStatus:PutServiceQuotaIncreaseRequestIntoTemplateResponse'

-> PutServiceQuotaIncreaseRequestIntoTemplateResponse

Create a value of PutServiceQuotaIncreaseRequestIntoTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceQuotaIncreaseRequestInTemplate:PutServiceQuotaIncreaseRequestIntoTemplateResponse', putServiceQuotaIncreaseRequestIntoTemplateResponse_serviceQuotaIncreaseRequestInTemplate - Information about the quota increase request.

$sel:httpStatus:PutServiceQuotaIncreaseRequestIntoTemplateResponse', putServiceQuotaIncreaseRequestIntoTemplateResponse_httpStatus - The response's http status code.

RequestServiceQuotaIncrease

data RequestServiceQuotaIncrease

See: newRequestServiceQuotaIncrease smart constructor.

Constructors

RequestServiceQuotaIncrease' Text Text Double

newRequestServiceQuotaIncrease

Arguments

:: Text
$sel:serviceCode:RequestServiceQuotaIncrease'

-> Text
$sel:quotaCode:RequestServiceQuotaIncrease'

-> Double
$sel:desiredValue:RequestServiceQuotaIncrease'

-> RequestServiceQuotaIncrease

Create a value of RequestServiceQuotaIncrease with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceCode:RequestServiceQuotaIncrease', requestServiceQuotaIncrease_serviceCode - The service identifier.

$sel:quotaCode:RequestServiceQuotaIncrease', requestServiceQuotaIncrease_quotaCode - The quota identifier.

$sel:desiredValue:RequestServiceQuotaIncrease', requestServiceQuotaIncrease_desiredValue - The new, increased value for the quota.

data RequestServiceQuotaIncreaseResponse

See: newRequestServiceQuotaIncreaseResponse smart constructor.

Constructors

RequestServiceQuotaIncreaseResponse' (Maybe RequestedServiceQuotaChange) Int

newRequestServiceQuotaIncreaseResponse

Arguments

:: Int
$sel:httpStatus:RequestServiceQuotaIncreaseResponse'

-> RequestServiceQuotaIncreaseResponse

Create a value of RequestServiceQuotaIncreaseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestedQuota:RequestServiceQuotaIncreaseResponse', requestServiceQuotaIncreaseResponse_requestedQuota - Information about the quota increase request.

$sel:httpStatus:RequestServiceQuotaIncreaseResponse', requestServiceQuotaIncreaseResponse_httpStatus - The response's http status code.

GetServiceQuotaIncreaseRequestFromTemplate

data GetServiceQuotaIncreaseRequestFromTemplate

See: newGetServiceQuotaIncreaseRequestFromTemplate smart constructor.

Constructors

GetServiceQuotaIncreaseRequestFromTemplate' Text Text Text

newGetServiceQuotaIncreaseRequestFromTemplate

Arguments

:: Text
$sel:serviceCode:GetServiceQuotaIncreaseRequestFromTemplate'

-> Text
$sel:quotaCode:GetServiceQuotaIncreaseRequestFromTemplate'

-> Text
$sel:awsRegion:GetServiceQuotaIncreaseRequestFromTemplate'

-> GetServiceQuotaIncreaseRequestFromTemplate

Create a value of GetServiceQuotaIncreaseRequestFromTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceCode:GetServiceQuotaIncreaseRequestFromTemplate', getServiceQuotaIncreaseRequestFromTemplate_serviceCode - The service identifier.

$sel:quotaCode:GetServiceQuotaIncreaseRequestFromTemplate', getServiceQuotaIncreaseRequestFromTemplate_quotaCode - The quota identifier.

$sel:awsRegion:GetServiceQuotaIncreaseRequestFromTemplate', getServiceQuotaIncreaseRequestFromTemplate_awsRegion - The AWS Region.

data GetServiceQuotaIncreaseRequestFromTemplateResponse

See: newGetServiceQuotaIncreaseRequestFromTemplateResponse smart constructor.

Constructors

GetServiceQuotaIncreaseRequestFromTemplateResponse' (Maybe ServiceQuotaIncreaseRequestInTemplate) Int

newGetServiceQuotaIncreaseRequestFromTemplateResponse

Arguments

:: Int
$sel:httpStatus:GetServiceQuotaIncreaseRequestFromTemplateResponse'

-> GetServiceQuotaIncreaseRequestFromTemplateResponse

Create a value of GetServiceQuotaIncreaseRequestFromTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceQuotaIncreaseRequestInTemplate:GetServiceQuotaIncreaseRequestFromTemplateResponse', getServiceQuotaIncreaseRequestFromTemplateResponse_serviceQuotaIncreaseRequestInTemplate - Information about the quota increase request.

$sel:httpStatus:GetServiceQuotaIncreaseRequestFromTemplateResponse', getServiceQuotaIncreaseRequestFromTemplateResponse_httpStatus - The response's http status code.

DisassociateServiceQuotaTemplate

data DisassociateServiceQuotaTemplate

See: newDisassociateServiceQuotaTemplate smart constructor.

Constructors

DisassociateServiceQuotaTemplate'

newDisassociateServiceQuotaTemplate :: DisassociateServiceQuotaTemplate

Create a value of DisassociateServiceQuotaTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

data DisassociateServiceQuotaTemplateResponse

See: newDisassociateServiceQuotaTemplateResponse smart constructor.

Constructors

DisassociateServiceQuotaTemplateResponse' Int

newDisassociateServiceQuotaTemplateResponse

Arguments

:: Int
$sel:httpStatus:DisassociateServiceQuotaTemplateResponse'

-> DisassociateServiceQuotaTemplateResponse

Create a value of DisassociateServiceQuotaTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateServiceQuotaTemplateResponse', disassociateServiceQuotaTemplateResponse_httpStatus - The response's http status code.

DeleteServiceQuotaIncreaseRequestFromTemplate

data DeleteServiceQuotaIncreaseRequestFromTemplate

See: newDeleteServiceQuotaIncreaseRequestFromTemplate smart constructor.

Constructors

DeleteServiceQuotaIncreaseRequestFromTemplate' Text Text Text

newDeleteServiceQuotaIncreaseRequestFromTemplate

Arguments

:: Text
$sel:serviceCode:DeleteServiceQuotaIncreaseRequestFromTemplate'

-> Text
$sel:quotaCode:DeleteServiceQuotaIncreaseRequestFromTemplate'

-> Text
$sel:awsRegion:DeleteServiceQuotaIncreaseRequestFromTemplate'

-> DeleteServiceQuotaIncreaseRequestFromTemplate

Create a value of DeleteServiceQuotaIncreaseRequestFromTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceCode:DeleteServiceQuotaIncreaseRequestFromTemplate', deleteServiceQuotaIncreaseRequestFromTemplate_serviceCode - The service identifier.

$sel:quotaCode:DeleteServiceQuotaIncreaseRequestFromTemplate', deleteServiceQuotaIncreaseRequestFromTemplate_quotaCode - The quota identifier.

$sel:awsRegion:DeleteServiceQuotaIncreaseRequestFromTemplate', deleteServiceQuotaIncreaseRequestFromTemplate_awsRegion - The AWS Region.

data DeleteServiceQuotaIncreaseRequestFromTemplateResponse

See: newDeleteServiceQuotaIncreaseRequestFromTemplateResponse smart constructor.

Constructors

DeleteServiceQuotaIncreaseRequestFromTemplateResponse' Int

newDeleteServiceQuotaIncreaseRequestFromTemplateResponse

Arguments

:: Int
$sel:httpStatus:DeleteServiceQuotaIncreaseRequestFromTemplateResponse'

-> DeleteServiceQuotaIncreaseRequestFromTemplateResponse

Create a value of DeleteServiceQuotaIncreaseRequestFromTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteServiceQuotaIncreaseRequestFromTemplateResponse', deleteServiceQuotaIncreaseRequestFromTemplateResponse_httpStatus - The response's http status code.

ListServiceQuotas (Paginated)

data ListServiceQuotas

See: newListServiceQuotas smart constructor.

Constructors

ListServiceQuotas' (Maybe Text) (Maybe Natural) Text

newListServiceQuotas

Arguments

:: Text
$sel:serviceCode:ListServiceQuotas'

-> ListServiceQuotas

Create a value of ListServiceQuotas with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListServiceQuotas', listServiceQuotas_nextToken - The token for the next page of results.

$sel:maxResults:ListServiceQuotas', listServiceQuotas_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

$sel:serviceCode:ListServiceQuotas', listServiceQuotas_serviceCode - The service identifier.

data ListServiceQuotasResponse

See: newListServiceQuotasResponse smart constructor.

Constructors

ListServiceQuotasResponse' (Maybe Text) (Maybe [ServiceQuota]) Int

newListServiceQuotasResponse

Arguments

:: Int
$sel:httpStatus:ListServiceQuotasResponse'

-> ListServiceQuotasResponse

Create a value of ListServiceQuotasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListServiceQuotas', listServiceQuotasResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:quotas:ListServiceQuotasResponse', listServiceQuotasResponse_quotas - Information about the quotas.

$sel:httpStatus:ListServiceQuotasResponse', listServiceQuotasResponse_httpStatus - The response's http status code.

ListRequestedServiceQuotaChangeHistory (Paginated)

data ListRequestedServiceQuotaChangeHistory

See: newListRequestedServiceQuotaChangeHistory smart constructor.

Constructors

ListRequestedServiceQuotaChangeHistory' (Maybe RequestStatus) (Maybe Text) (Maybe Text) (Maybe Natural)

newListRequestedServiceQuotaChangeHistory :: ListRequestedServiceQuotaChangeHistory

Create a value of ListRequestedServiceQuotaChangeHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListRequestedServiceQuotaChangeHistory', listRequestedServiceQuotaChangeHistory_status - The status of the quota increase request.

$sel:nextToken:ListRequestedServiceQuotaChangeHistory', listRequestedServiceQuotaChangeHistory_nextToken - The token for the next page of results.

$sel:serviceCode:ListRequestedServiceQuotaChangeHistory', listRequestedServiceQuotaChangeHistory_serviceCode - The service identifier.

$sel:maxResults:ListRequestedServiceQuotaChangeHistory', listRequestedServiceQuotaChangeHistory_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

data ListRequestedServiceQuotaChangeHistoryResponse

See: newListRequestedServiceQuotaChangeHistoryResponse smart constructor.

Constructors

ListRequestedServiceQuotaChangeHistoryResponse' (Maybe Text) (Maybe [RequestedServiceQuotaChange]) Int

newListRequestedServiceQuotaChangeHistoryResponse

Arguments

:: Int
$sel:httpStatus:ListRequestedServiceQuotaChangeHistoryResponse'

-> ListRequestedServiceQuotaChangeHistoryResponse

Create a value of ListRequestedServiceQuotaChangeHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRequestedServiceQuotaChangeHistory', listRequestedServiceQuotaChangeHistoryResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:requestedQuotas:ListRequestedServiceQuotaChangeHistoryResponse', listRequestedServiceQuotaChangeHistoryResponse_requestedQuotas - Information about the quota increase requests.

$sel:httpStatus:ListRequestedServiceQuotaChangeHistoryResponse', listRequestedServiceQuotaChangeHistoryResponse_httpStatus - The response's http status code.

ListServiceQuotaIncreaseRequestsInTemplate (Paginated)

data ListServiceQuotaIncreaseRequestsInTemplate

See: newListServiceQuotaIncreaseRequestsInTemplate smart constructor.

Constructors

ListServiceQuotaIncreaseRequestsInTemplate' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Natural)

newListServiceQuotaIncreaseRequestsInTemplate :: ListServiceQuotaIncreaseRequestsInTemplate

Create a value of ListServiceQuotaIncreaseRequestsInTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListServiceQuotaIncreaseRequestsInTemplate', listServiceQuotaIncreaseRequestsInTemplate_nextToken - The token for the next page of results.

$sel:awsRegion:ListServiceQuotaIncreaseRequestsInTemplate', listServiceQuotaIncreaseRequestsInTemplate_awsRegion - The AWS Region.

$sel:serviceCode:ListServiceQuotaIncreaseRequestsInTemplate', listServiceQuotaIncreaseRequestsInTemplate_serviceCode - The service identifier.

$sel:maxResults:ListServiceQuotaIncreaseRequestsInTemplate', listServiceQuotaIncreaseRequestsInTemplate_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

data ListServiceQuotaIncreaseRequestsInTemplateResponse

See: newListServiceQuotaIncreaseRequestsInTemplateResponse smart constructor.

Constructors

ListServiceQuotaIncreaseRequestsInTemplateResponse' (Maybe [ServiceQuotaIncreaseRequestInTemplate]) (Maybe Text) Int

newListServiceQuotaIncreaseRequestsInTemplateResponse

Arguments

:: Int
$sel:httpStatus:ListServiceQuotaIncreaseRequestsInTemplateResponse'

-> ListServiceQuotaIncreaseRequestsInTemplateResponse

Create a value of ListServiceQuotaIncreaseRequestsInTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceQuotaIncreaseRequestInTemplateList:ListServiceQuotaIncreaseRequestsInTemplateResponse', listServiceQuotaIncreaseRequestsInTemplateResponse_serviceQuotaIncreaseRequestInTemplateList - Information about the quota increase requests.

$sel:nextToken:ListServiceQuotaIncreaseRequestsInTemplate', listServiceQuotaIncreaseRequestsInTemplateResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListServiceQuotaIncreaseRequestsInTemplateResponse', listServiceQuotaIncreaseRequestsInTemplateResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (NonEmpty Tag)

newTagResource

Arguments

:: Text
$sel:resourceARN:TagResource'

-> NonEmpty Tag
$sel:tags:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARN:TagResource', tagResource_resourceARN - The Amazon Resource Name (ARN) for the applied quota. You can get this information by using the Service Quotas console, or by listing the quotas using the list-service-quotas AWS CLI command or the ListServiceQuotas AWS API operation.

$sel:tags:TagResource', tagResource_tags - The tags that you want to add to the resource.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

ListRequestedServiceQuotaChangeHistoryByQuota (Paginated)

data ListRequestedServiceQuotaChangeHistoryByQuota

See: newListRequestedServiceQuotaChangeHistoryByQuota smart constructor.

Constructors

ListRequestedServiceQuotaChangeHistoryByQuota' (Maybe RequestStatus) (Maybe Text) (Maybe Natural) Text Text

newListRequestedServiceQuotaChangeHistoryByQuota

Arguments

:: Text
$sel:serviceCode:ListRequestedServiceQuotaChangeHistoryByQuota'

-> Text
$sel:quotaCode:ListRequestedServiceQuotaChangeHistoryByQuota'

-> ListRequestedServiceQuotaChangeHistoryByQuota

Create a value of ListRequestedServiceQuotaChangeHistoryByQuota with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListRequestedServiceQuotaChangeHistoryByQuota', listRequestedServiceQuotaChangeHistoryByQuota_status - The status value of the quota increase request.

$sel:nextToken:ListRequestedServiceQuotaChangeHistoryByQuota', listRequestedServiceQuotaChangeHistoryByQuota_nextToken - The token for the next page of results.

$sel:maxResults:ListRequestedServiceQuotaChangeHistoryByQuota', listRequestedServiceQuotaChangeHistoryByQuota_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

$sel:serviceCode:ListRequestedServiceQuotaChangeHistoryByQuota', listRequestedServiceQuotaChangeHistoryByQuota_serviceCode - The service identifier.

$sel:quotaCode:ListRequestedServiceQuotaChangeHistoryByQuota', listRequestedServiceQuotaChangeHistoryByQuota_quotaCode - The quota identifier.

data ListRequestedServiceQuotaChangeHistoryByQuotaResponse

See: newListRequestedServiceQuotaChangeHistoryByQuotaResponse smart constructor.

Constructors

ListRequestedServiceQuotaChangeHistoryByQuotaResponse' (Maybe Text) (Maybe [RequestedServiceQuotaChange]) Int

newListRequestedServiceQuotaChangeHistoryByQuotaResponse

Arguments

:: Int
$sel:httpStatus:ListRequestedServiceQuotaChangeHistoryByQuotaResponse'

-> ListRequestedServiceQuotaChangeHistoryByQuotaResponse

Create a value of ListRequestedServiceQuotaChangeHistoryByQuotaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRequestedServiceQuotaChangeHistoryByQuota', listRequestedServiceQuotaChangeHistoryByQuotaResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:requestedQuotas:ListRequestedServiceQuotaChangeHistoryByQuotaResponse', listRequestedServiceQuotaChangeHistoryByQuotaResponse_requestedQuotas - Information about the quota increase requests.

$sel:httpStatus:ListRequestedServiceQuotaChangeHistoryByQuotaResponse', listRequestedServiceQuotaChangeHistoryByQuotaResponse_httpStatus - The response's http status code.

GetRequestedServiceQuotaChange

data GetRequestedServiceQuotaChange

See: newGetRequestedServiceQuotaChange smart constructor.

Constructors

GetRequestedServiceQuotaChange' Text

newGetRequestedServiceQuotaChange

Arguments

:: Text
$sel:requestId:GetRequestedServiceQuotaChange'

-> GetRequestedServiceQuotaChange

Create a value of GetRequestedServiceQuotaChange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:GetRequestedServiceQuotaChange', getRequestedServiceQuotaChange_requestId - The ID of the quota increase request.

data GetRequestedServiceQuotaChangeResponse

See: newGetRequestedServiceQuotaChangeResponse smart constructor.

Constructors

GetRequestedServiceQuotaChangeResponse' (Maybe RequestedServiceQuotaChange) Int

newGetRequestedServiceQuotaChangeResponse

Arguments

:: Int
$sel:httpStatus:GetRequestedServiceQuotaChangeResponse'

-> GetRequestedServiceQuotaChangeResponse

Create a value of GetRequestedServiceQuotaChangeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestedQuota:GetRequestedServiceQuotaChangeResponse', getRequestedServiceQuotaChangeResponse_requestedQuota - Information about the quota increase request.

$sel:httpStatus:GetRequestedServiceQuotaChangeResponse', getRequestedServiceQuotaChangeResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceARN:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARN:UntagResource', untagResource_resourceARN - The Amazon Resource Name (ARN) for the applied quota that you want to untag. You can get this information by using the Service Quotas console, or by listing the quotas using the list-service-quotas AWS CLI command or the ListServiceQuotas AWS API operation.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The keys of the tags that you want to remove from the resource.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

GetAWSDefaultServiceQuota

data GetAWSDefaultServiceQuota

See: newGetAWSDefaultServiceQuota smart constructor.

Constructors

GetAWSDefaultServiceQuota' Text Text

newGetAWSDefaultServiceQuota

Arguments

:: Text
$sel:serviceCode:GetAWSDefaultServiceQuota'

-> Text
$sel:quotaCode:GetAWSDefaultServiceQuota'

-> GetAWSDefaultServiceQuota

Create a value of GetAWSDefaultServiceQuota with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceCode:GetAWSDefaultServiceQuota', getAWSDefaultServiceQuota_serviceCode - The service identifier.

$sel:quotaCode:GetAWSDefaultServiceQuota', getAWSDefaultServiceQuota_quotaCode - The quota identifier.

data GetAWSDefaultServiceQuotaResponse

See: newGetAWSDefaultServiceQuotaResponse smart constructor.

Constructors

GetAWSDefaultServiceQuotaResponse' (Maybe ServiceQuota) Int

newGetAWSDefaultServiceQuotaResponse

Arguments

:: Int
$sel:httpStatus:GetAWSDefaultServiceQuotaResponse'

-> GetAWSDefaultServiceQuotaResponse

Create a value of GetAWSDefaultServiceQuotaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quota:GetAWSDefaultServiceQuotaResponse', getAWSDefaultServiceQuotaResponse_quota - Information about the quota.

$sel:httpStatus:GetAWSDefaultServiceQuotaResponse', getAWSDefaultServiceQuotaResponse_httpStatus - The response's http status code.

Types

ErrorCode

newtype ErrorCode

Constructors

ErrorCode'

Fields

fromErrorCode :: Text

Bundled Patterns

pattern ErrorCode_DEPENDENCY_ACCESS_DENIED_ERROR :: ErrorCode

pattern ErrorCode_DEPENDENCY_SERVICE_ERROR :: ErrorCode

pattern ErrorCode_DEPENDENCY_THROTTLING_ERROR :: ErrorCode

pattern ErrorCode_SERVICE_QUOTA_NOT_AVAILABLE_ERROR :: ErrorCode

PeriodUnit

newtype PeriodUnit

Constructors

PeriodUnit'

Fields

fromPeriodUnit :: Text

Bundled Patterns

pattern PeriodUnit_DAY :: PeriodUnit

pattern PeriodUnit_HOUR :: PeriodUnit

pattern PeriodUnit_MICROSECOND :: PeriodUnit

pattern PeriodUnit_MILLISECOND :: PeriodUnit

pattern PeriodUnit_MINUTE :: PeriodUnit

pattern PeriodUnit_SECOND :: PeriodUnit

pattern PeriodUnit_WEEK :: PeriodUnit

RequestStatus

newtype RequestStatus

Constructors

RequestStatus'

Fields

fromRequestStatus :: Text

Bundled Patterns

pattern RequestStatus_APPROVED :: RequestStatus

pattern RequestStatus_CASE_CLOSED :: RequestStatus

pattern RequestStatus_CASE_OPENED :: RequestStatus

pattern RequestStatus_DENIED :: RequestStatus

pattern RequestStatus_PENDING :: RequestStatus

ServiceQuotaTemplateAssociationStatus

newtype ServiceQuotaTemplateAssociationStatus

Constructors

ServiceQuotaTemplateAssociationStatus'

Fields

fromServiceQuotaTemplateAssociationStatus :: Text

Bundled Patterns

pattern ServiceQuotaTemplateAssociationStatus_ASSOCIATED :: ServiceQuotaTemplateAssociationStatus

pattern ServiceQuotaTemplateAssociationStatus_DISASSOCIATED :: ServiceQuotaTemplateAssociationStatus

ErrorReason

data ErrorReason

An error that explains why an action did not succeed.

See: newErrorReason smart constructor.

Constructors

ErrorReason' (Maybe ErrorCode) (Maybe Text)

newErrorReason :: ErrorReason

Create a value of ErrorReason with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:ErrorReason', errorReason_errorCode - Service Quotas returns the following error values:

DEPENDENCY_ACCESS_DENIED_ERROR - The caller does not have the required permissions to complete the action. To resolve the error, you must have permission to access the service or quota.

DEPENDENCY_THROTTLING_ERROR - The service is throttling Service Quotas.

DEPENDENCY_SERVICE_ERROR - The service is not available.

SERVICE_QUOTA_NOT_AVAILABLE_ERROR - There was an error in Service Quotas.

$sel:errorMessage:ErrorReason', errorReason_errorMessage - The error message.

MetricInfo

data MetricInfo

Information about the CloudWatch metric that reflects quota usage.

See: newMetricInfo smart constructor.

Constructors

MetricInfo' (Maybe (HashMap Text Text)) (Maybe Text) (Maybe Text) (Maybe Text)

newMetricInfo :: MetricInfo

Create a value of MetricInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricDimensions:MetricInfo', metricInfo_metricDimensions - The metric dimension. This is a name/value pair that is part of the identity of a metric.

$sel:metricName:MetricInfo', metricInfo_metricName - The name of the metric.

$sel:metricStatisticRecommendation:MetricInfo', metricInfo_metricStatisticRecommendation - The metric statistic that we recommend you use when determining quota usage.

$sel:metricNamespace:MetricInfo', metricInfo_metricNamespace - The namespace of the metric.

QuotaPeriod

data QuotaPeriod

Information about the quota period.

See: newQuotaPeriod smart constructor.

Constructors

QuotaPeriod' (Maybe PeriodUnit) (Maybe Int)

newQuotaPeriod :: QuotaPeriod

Create a value of QuotaPeriod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:periodUnit:QuotaPeriod', quotaPeriod_periodUnit - The time unit.

$sel:periodValue:QuotaPeriod', quotaPeriod_periodValue - The value.

RequestedServiceQuotaChange

data RequestedServiceQuotaChange

Information about a quota increase request.

See: newRequestedServiceQuotaChange smart constructor.

Constructors

RequestedServiceQuotaChange' (Maybe RequestStatus) (Maybe POSIX) (Maybe Bool) (Maybe POSIX) (Maybe Double) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text)

newRequestedServiceQuotaChange :: RequestedServiceQuotaChange

Create a value of RequestedServiceQuotaChange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RequestedServiceQuotaChange', requestedServiceQuotaChange_status - The state of the quota increase request.

$sel:lastUpdated:RequestedServiceQuotaChange', requestedServiceQuotaChange_lastUpdated - The date and time of the most recent change.

$sel:globalQuota:RequestedServiceQuotaChange', requestedServiceQuotaChange_globalQuota - Indicates whether the quota is global.

$sel:created:RequestedServiceQuotaChange', requestedServiceQuotaChange_created - The date and time when the quota increase request was received and the case ID was created.

$sel:desiredValue:RequestedServiceQuotaChange', requestedServiceQuotaChange_desiredValue - The new, increased value for the quota.

$sel:quotaArn:RequestedServiceQuotaChange', requestedServiceQuotaChange_quotaArn - The Amazon Resource Name (ARN) of the quota.

$sel:caseId:RequestedServiceQuotaChange', requestedServiceQuotaChange_caseId - The case ID.

$sel:serviceName:RequestedServiceQuotaChange', requestedServiceQuotaChange_serviceName - The service name.

$sel:id:RequestedServiceQuotaChange', requestedServiceQuotaChange_id - The unique identifier.

$sel:serviceCode:RequestedServiceQuotaChange', requestedServiceQuotaChange_serviceCode - The service identifier.

$sel:quotaCode:RequestedServiceQuotaChange', requestedServiceQuotaChange_quotaCode - The quota identifier.

$sel:unit:RequestedServiceQuotaChange', requestedServiceQuotaChange_unit - The unit of measurement.

$sel:requester:RequestedServiceQuotaChange', requestedServiceQuotaChange_requester - The IAM identity of the requester.

$sel:quotaName:RequestedServiceQuotaChange', requestedServiceQuotaChange_quotaName - The quota name.

ServiceInfo

data ServiceInfo

Information about a service.

See: newServiceInfo smart constructor.

Constructors

ServiceInfo' (Maybe Text) (Maybe Text)

newServiceInfo :: ServiceInfo

Create a value of ServiceInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceName:ServiceInfo', serviceInfo_serviceName - The service name.

$sel:serviceCode:ServiceInfo', serviceInfo_serviceCode - The service identifier.

ServiceQuota

data ServiceQuota

Information about a quota.

See: newServiceQuota smart constructor.

Constructors

ServiceQuota' (Maybe Bool) (Maybe QuotaPeriod) (Maybe Double) (Maybe Text) (Maybe MetricInfo) (Maybe ErrorReason) (Maybe Bool) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text)

newServiceQuota :: ServiceQuota

Create a value of ServiceQuota with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalQuota:ServiceQuota', serviceQuota_globalQuota - Indicates whether the quota is global.

$sel:period:ServiceQuota', serviceQuota_period - The period of time.

$sel:value:ServiceQuota', serviceQuota_value - The quota value.

$sel:quotaArn:ServiceQuota', serviceQuota_quotaArn - The Amazon Resource Name (ARN) of the quota.

$sel:usageMetric:ServiceQuota', serviceQuota_usageMetric - Information about the measurement.

$sel:errorReason:ServiceQuota', serviceQuota_errorReason - The error code and error reason.

$sel:adjustable:ServiceQuota', serviceQuota_adjustable - Indicates whether the quota value can be increased.

$sel:serviceName:ServiceQuota', serviceQuota_serviceName - The service name.

$sel:serviceCode:ServiceQuota', serviceQuota_serviceCode - The service identifier.

$sel:quotaCode:ServiceQuota', serviceQuota_quotaCode - The quota identifier.

$sel:unit:ServiceQuota', serviceQuota_unit - The unit of measurement.

$sel:quotaName:ServiceQuota', serviceQuota_quotaName - The quota name.

ServiceQuotaIncreaseRequestInTemplate

data ServiceQuotaIncreaseRequestInTemplate

Information about a quota increase request.

See: newServiceQuotaIncreaseRequestInTemplate smart constructor.

Constructors

ServiceQuotaIncreaseRequestInTemplate' (Maybe Bool) (Maybe Double) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text)

newServiceQuotaIncreaseRequestInTemplate :: ServiceQuotaIncreaseRequestInTemplate

Create a value of ServiceQuotaIncreaseRequestInTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalQuota:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_globalQuota - Indicates whether the quota is global.

$sel:desiredValue:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_desiredValue - The new, increased value of the quota.

$sel:serviceName:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_serviceName - The service name.

$sel:awsRegion:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_awsRegion - The AWS Region.

$sel:serviceCode:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_serviceCode - The service identifier.

$sel:quotaCode:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_quotaCode - The quota identifier.

$sel:unit:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_unit - The unit of measurement.

$sel:quotaName:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_quotaName - The quota name.

Tag

data Tag

A complex data type that contains a tag key and tag value.

See: newTag smart constructor.

Constructors

Tag' Text Text

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - A string that contains a tag key. The string length should be between 1 and 128 characters. Valid characters include a-z, A-Z, 0-9, space, and the special characters _ - . : / = + @.

$sel:value:Tag', tag_value - A string that contains an optional tag value. The string length should be between 0 and 256 characters. Valid characters include a-z, A-Z, 0-9, space, and the special characters _ - . : / = + @.