Amazonka.ServiceQuotas.ListRequestedServiceQuotaChangeHistoryByQuota

Description

Retrieves the quota increase requests for the specified quota.

This operation returns paginated results.

Creating a Request

data ListRequestedServiceQuotaChangeHistoryByQuota

See: newListRequestedServiceQuotaChangeHistoryByQuota smart constructor.

Constructors

ListRequestedServiceQuotaChangeHistoryByQuota'

Fields

status :: Maybe RequestStatus

The status value of the quota increase request.

nextToken :: Maybe Text

The token for the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

serviceCode :: Text

The service identifier.

quotaCode :: Text

The quota identifier.

newListRequestedServiceQuotaChangeHistoryByQuota

Arguments

:: Text
$sel:serviceCode:ListRequestedServiceQuotaChangeHistoryByQuota'

-> Text
$sel:quotaCode:ListRequestedServiceQuotaChangeHistoryByQuota'

-> ListRequestedServiceQuotaChangeHistoryByQuota

Create a value of ListRequestedServiceQuotaChangeHistoryByQuota with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListRequestedServiceQuotaChangeHistoryByQuota', listRequestedServiceQuotaChangeHistoryByQuota_status - The status value of the quota increase request.

$sel:nextToken:ListRequestedServiceQuotaChangeHistoryByQuota', listRequestedServiceQuotaChangeHistoryByQuota_nextToken - The token for the next page of results.

$sel:maxResults:ListRequestedServiceQuotaChangeHistoryByQuota', listRequestedServiceQuotaChangeHistoryByQuota_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

$sel:serviceCode:ListRequestedServiceQuotaChangeHistoryByQuota', listRequestedServiceQuotaChangeHistoryByQuota_serviceCode - The service identifier.

$sel:quotaCode:ListRequestedServiceQuotaChangeHistoryByQuota', listRequestedServiceQuotaChangeHistoryByQuota_quotaCode - The quota identifier.

Request Lenses

listRequestedServiceQuotaChangeHistoryByQuota_status :: Lens' ListRequestedServiceQuotaChangeHistoryByQuota (Maybe RequestStatus)

The status value of the quota increase request.

listRequestedServiceQuotaChangeHistoryByQuota_nextToken :: Lens' ListRequestedServiceQuotaChangeHistoryByQuota (Maybe Text)

The token for the next page of results.

listRequestedServiceQuotaChangeHistoryByQuota_maxResults :: Lens' ListRequestedServiceQuotaChangeHistoryByQuota (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

listRequestedServiceQuotaChangeHistoryByQuota_serviceCode :: Lens' ListRequestedServiceQuotaChangeHistoryByQuota Text

The service identifier.

listRequestedServiceQuotaChangeHistoryByQuota_quotaCode :: Lens' ListRequestedServiceQuotaChangeHistoryByQuota Text

The quota identifier.

Destructuring the Response

data ListRequestedServiceQuotaChangeHistoryByQuotaResponse

See: newListRequestedServiceQuotaChangeHistoryByQuotaResponse smart constructor.

Constructors

ListRequestedServiceQuotaChangeHistoryByQuotaResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

requestedQuotas :: Maybe [RequestedServiceQuotaChange]

Information about the quota increase requests.

httpStatus :: Int

The response's http status code.

newListRequestedServiceQuotaChangeHistoryByQuotaResponse

Arguments

:: Int
$sel:httpStatus:ListRequestedServiceQuotaChangeHistoryByQuotaResponse'

-> ListRequestedServiceQuotaChangeHistoryByQuotaResponse

Create a value of ListRequestedServiceQuotaChangeHistoryByQuotaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRequestedServiceQuotaChangeHistoryByQuota', listRequestedServiceQuotaChangeHistoryByQuotaResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:requestedQuotas:ListRequestedServiceQuotaChangeHistoryByQuotaResponse', listRequestedServiceQuotaChangeHistoryByQuotaResponse_requestedQuotas - Information about the quota increase requests.

$sel:httpStatus:ListRequestedServiceQuotaChangeHistoryByQuotaResponse', listRequestedServiceQuotaChangeHistoryByQuotaResponse_httpStatus - The response's http status code.

Response Lenses

listRequestedServiceQuotaChangeHistoryByQuotaResponse_nextToken :: Lens' ListRequestedServiceQuotaChangeHistoryByQuotaResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listRequestedServiceQuotaChangeHistoryByQuotaResponse_requestedQuotas :: Lens' ListRequestedServiceQuotaChangeHistoryByQuotaResponse (Maybe [RequestedServiceQuotaChange])

Information about the quota increase requests.

listRequestedServiceQuotaChangeHistoryByQuotaResponse_httpStatus :: Lens' ListRequestedServiceQuotaChangeHistoryByQuotaResponse Int

The response's http status code.