Amazonka.ServiceQuotas.ListServiceQuotaIncreaseRequestsInTemplate

Description

Lists the quota increase requests in the specified quota request template.

This operation returns paginated results.

Creating a Request

data ListServiceQuotaIncreaseRequestsInTemplate

See: newListServiceQuotaIncreaseRequestsInTemplate smart constructor.

Constructors

ListServiceQuotaIncreaseRequestsInTemplate'

Fields

nextToken :: Maybe Text

The token for the next page of results.

awsRegion :: Maybe Text

The AWS Region.

serviceCode :: Maybe Text

The service identifier.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

newListServiceQuotaIncreaseRequestsInTemplate :: ListServiceQuotaIncreaseRequestsInTemplate

Create a value of ListServiceQuotaIncreaseRequestsInTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListServiceQuotaIncreaseRequestsInTemplate', listServiceQuotaIncreaseRequestsInTemplate_nextToken - The token for the next page of results.

$sel:awsRegion:ListServiceQuotaIncreaseRequestsInTemplate', listServiceQuotaIncreaseRequestsInTemplate_awsRegion - The AWS Region.

$sel:serviceCode:ListServiceQuotaIncreaseRequestsInTemplate', listServiceQuotaIncreaseRequestsInTemplate_serviceCode - The service identifier.

$sel:maxResults:ListServiceQuotaIncreaseRequestsInTemplate', listServiceQuotaIncreaseRequestsInTemplate_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

Request Lenses

listServiceQuotaIncreaseRequestsInTemplate_nextToken :: Lens' ListServiceQuotaIncreaseRequestsInTemplate (Maybe Text)

The token for the next page of results.

listServiceQuotaIncreaseRequestsInTemplate_awsRegion :: Lens' ListServiceQuotaIncreaseRequestsInTemplate (Maybe Text)

The AWS Region.

listServiceQuotaIncreaseRequestsInTemplate_serviceCode :: Lens' ListServiceQuotaIncreaseRequestsInTemplate (Maybe Text)

The service identifier.

listServiceQuotaIncreaseRequestsInTemplate_maxResults :: Lens' ListServiceQuotaIncreaseRequestsInTemplate (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

Destructuring the Response

data ListServiceQuotaIncreaseRequestsInTemplateResponse

See: newListServiceQuotaIncreaseRequestsInTemplateResponse smart constructor.

Constructors

ListServiceQuotaIncreaseRequestsInTemplateResponse'

Fields

serviceQuotaIncreaseRequestInTemplateList :: Maybe [ServiceQuotaIncreaseRequestInTemplate]

Information about the quota increase requests.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newListServiceQuotaIncreaseRequestsInTemplateResponse

Arguments

:: Int
$sel:httpStatus:ListServiceQuotaIncreaseRequestsInTemplateResponse'

-> ListServiceQuotaIncreaseRequestsInTemplateResponse

Create a value of ListServiceQuotaIncreaseRequestsInTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceQuotaIncreaseRequestInTemplateList:ListServiceQuotaIncreaseRequestsInTemplateResponse', listServiceQuotaIncreaseRequestsInTemplateResponse_serviceQuotaIncreaseRequestInTemplateList - Information about the quota increase requests.

$sel:nextToken:ListServiceQuotaIncreaseRequestsInTemplate', listServiceQuotaIncreaseRequestsInTemplateResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListServiceQuotaIncreaseRequestsInTemplateResponse', listServiceQuotaIncreaseRequestsInTemplateResponse_httpStatus - The response's http status code.

Response Lenses

listServiceQuotaIncreaseRequestsInTemplateResponse_serviceQuotaIncreaseRequestInTemplateList :: Lens' ListServiceQuotaIncreaseRequestsInTemplateResponse (Maybe [ServiceQuotaIncreaseRequestInTemplate])

Information about the quota increase requests.

listServiceQuotaIncreaseRequestsInTemplateResponse_nextToken :: Lens' ListServiceQuotaIncreaseRequestsInTemplateResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listServiceQuotaIncreaseRequestsInTemplateResponse_httpStatus :: Lens' ListServiceQuotaIncreaseRequestsInTemplateResponse Int

The response's http status code.