Amazonka.ServiceQuotas.Types

Description

Service Configuration

defaultService :: Service

API version 2019-06-24 of the Amazon Quotas SDK configuration.

Errors

_TagPolicyViolationException :: AsError a => Getting (First ServiceError) a ServiceError

The specified tag is a reserved word and cannot be used.

_NoAvailableOrganizationException :: AsError a => Getting (First ServiceError) a ServiceError

The account making this call is not a member of an organization.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient permission to perform this action.

_TemplatesNotAvailableInRegionException :: AsError a => Getting (First ServiceError) a ServiceError

The Service Quotas template is not available in this AWS Region.

_DependencyAccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You can't perform this action because a dependency does not have access.

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource already exists.

_AWSServiceAccessNotEnabledException :: AsError a => Getting (First ServiceError) a ServiceError

The action you attempted is not allowed unless Service Access with Service Quotas is enabled in your organization.

_InvalidResourceStateException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is in an invalid state.

_TooManyTagsException :: AsError a => Getting (First ServiceError) a ServiceError

You've exceeded the number of tags allowed for a resource. For more information, see Tag restrictions in the Service Quotas User Guide.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Due to throttling, the request was denied. Slow down the rate of request calls, or request an increase for this quota.

_QuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded your service quota. To perform the requested action, remove some of the relevant resources, or use Service Quotas to request a service quota increase.

_ServiceException :: AsError a => Getting (First ServiceError) a ServiceError

Something went wrong.

_IllegalArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

Invalid input was provided.

_ServiceQuotaTemplateNotInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The quota request template is not associated with your organization.

_OrganizationNotInAllFeaturesModeException :: AsError a => Getting (First ServiceError) a ServiceError

The organization that your account belongs to is not in All Features mode.

_InvalidPaginationTokenException :: AsError a => Getting (First ServiceError) a ServiceError

Invalid input was provided.

_NoSuchResourceException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

ErrorCode

newtype ErrorCode

Constructors

ErrorCode'

Fields

fromErrorCode :: Text

Bundled Patterns

pattern ErrorCode_DEPENDENCY_ACCESS_DENIED_ERROR :: ErrorCode

pattern ErrorCode_DEPENDENCY_SERVICE_ERROR :: ErrorCode

pattern ErrorCode_DEPENDENCY_THROTTLING_ERROR :: ErrorCode

pattern ErrorCode_SERVICE_QUOTA_NOT_AVAILABLE_ERROR :: ErrorCode

PeriodUnit

newtype PeriodUnit

Constructors

PeriodUnit'

Fields

fromPeriodUnit :: Text

Bundled Patterns

pattern PeriodUnit_DAY :: PeriodUnit

pattern PeriodUnit_HOUR :: PeriodUnit

pattern PeriodUnit_MICROSECOND :: PeriodUnit

pattern PeriodUnit_MILLISECOND :: PeriodUnit

pattern PeriodUnit_MINUTE :: PeriodUnit

pattern PeriodUnit_SECOND :: PeriodUnit

pattern PeriodUnit_WEEK :: PeriodUnit

RequestStatus

newtype RequestStatus

Constructors

RequestStatus'

Fields

fromRequestStatus :: Text

Bundled Patterns

pattern RequestStatus_APPROVED :: RequestStatus

pattern RequestStatus_CASE_CLOSED :: RequestStatus

pattern RequestStatus_CASE_OPENED :: RequestStatus

pattern RequestStatus_DENIED :: RequestStatus

pattern RequestStatus_PENDING :: RequestStatus

ServiceQuotaTemplateAssociationStatus

newtype ServiceQuotaTemplateAssociationStatus

Constructors

ServiceQuotaTemplateAssociationStatus'

Fields

fromServiceQuotaTemplateAssociationStatus :: Text

Bundled Patterns

pattern ServiceQuotaTemplateAssociationStatus_ASSOCIATED :: ServiceQuotaTemplateAssociationStatus

pattern ServiceQuotaTemplateAssociationStatus_DISASSOCIATED :: ServiceQuotaTemplateAssociationStatus

ErrorReason

data ErrorReason

An error that explains why an action did not succeed.

See: newErrorReason smart constructor.

Constructors

ErrorReason'

Fields

errorCode :: Maybe ErrorCode

Service Quotas returns the following error values:

DEPENDENCY_ACCESS_DENIED_ERROR - The caller does not have the required permissions to complete the action. To resolve the error, you must have permission to access the service or quota.

DEPENDENCY_THROTTLING_ERROR - The service is throttling Service Quotas.

DEPENDENCY_SERVICE_ERROR - The service is not available.

SERVICE_QUOTA_NOT_AVAILABLE_ERROR - There was an error in Service Quotas.

errorMessage :: Maybe Text

The error message.

newErrorReason :: ErrorReason

Create a value of ErrorReason with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:ErrorReason', errorReason_errorCode - Service Quotas returns the following error values:

DEPENDENCY_ACCESS_DENIED_ERROR - The caller does not have the required permissions to complete the action. To resolve the error, you must have permission to access the service or quota.

DEPENDENCY_THROTTLING_ERROR - The service is throttling Service Quotas.

DEPENDENCY_SERVICE_ERROR - The service is not available.

SERVICE_QUOTA_NOT_AVAILABLE_ERROR - There was an error in Service Quotas.

$sel:errorMessage:ErrorReason', errorReason_errorMessage - The error message.

errorReason_errorCode :: Lens' ErrorReason (Maybe ErrorCode)

Service Quotas returns the following error values:

DEPENDENCY_ACCESS_DENIED_ERROR - The caller does not have the required permissions to complete the action. To resolve the error, you must have permission to access the service or quota.

DEPENDENCY_THROTTLING_ERROR - The service is throttling Service Quotas.

DEPENDENCY_SERVICE_ERROR - The service is not available.

SERVICE_QUOTA_NOT_AVAILABLE_ERROR - There was an error in Service Quotas.

errorReason_errorMessage :: Lens' ErrorReason (Maybe Text)

The error message.

MetricInfo

data MetricInfo

Information about the CloudWatch metric that reflects quota usage.

See: newMetricInfo smart constructor.

Constructors

MetricInfo'

Fields

metricDimensions :: Maybe (HashMap Text Text)

The metric dimension. This is a name/value pair that is part of the identity of a metric.

metricName :: Maybe Text

The name of the metric.

metricStatisticRecommendation :: Maybe Text

The metric statistic that we recommend you use when determining quota usage.

metricNamespace :: Maybe Text

The namespace of the metric.

newMetricInfo :: MetricInfo

Create a value of MetricInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricDimensions:MetricInfo', metricInfo_metricDimensions - The metric dimension. This is a name/value pair that is part of the identity of a metric.

$sel:metricName:MetricInfo', metricInfo_metricName - The name of the metric.

$sel:metricStatisticRecommendation:MetricInfo', metricInfo_metricStatisticRecommendation - The metric statistic that we recommend you use when determining quota usage.

$sel:metricNamespace:MetricInfo', metricInfo_metricNamespace - The namespace of the metric.

metricInfo_metricDimensions :: Lens' MetricInfo (Maybe (HashMap Text Text))

The metric dimension. This is a name/value pair that is part of the identity of a metric.

metricInfo_metricName :: Lens' MetricInfo (Maybe Text)

The name of the metric.

metricInfo_metricStatisticRecommendation :: Lens' MetricInfo (Maybe Text)

The metric statistic that we recommend you use when determining quota usage.

metricInfo_metricNamespace :: Lens' MetricInfo (Maybe Text)

The namespace of the metric.

QuotaPeriod

data QuotaPeriod

Information about the quota period.

See: newQuotaPeriod smart constructor.

Constructors

QuotaPeriod'

Fields

periodUnit :: Maybe PeriodUnit

The time unit.

periodValue :: Maybe Int

The value.

newQuotaPeriod :: QuotaPeriod

Create a value of QuotaPeriod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:periodUnit:QuotaPeriod', quotaPeriod_periodUnit - The time unit.

$sel:periodValue:QuotaPeriod', quotaPeriod_periodValue - The value.

quotaPeriod_periodUnit :: Lens' QuotaPeriod (Maybe PeriodUnit)

The time unit.

quotaPeriod_periodValue :: Lens' QuotaPeriod (Maybe Int)

The value.

RequestedServiceQuotaChange

data RequestedServiceQuotaChange

Information about a quota increase request.

See: newRequestedServiceQuotaChange smart constructor.

Constructors

RequestedServiceQuotaChange'

Fields

status :: Maybe RequestStatus

The state of the quota increase request.

lastUpdated :: Maybe POSIX

The date and time of the most recent change.

globalQuota :: Maybe Bool

Indicates whether the quota is global.

created :: Maybe POSIX

The date and time when the quota increase request was received and the case ID was created.

desiredValue :: Maybe Double

The new, increased value for the quota.

quotaArn :: Maybe Text

The Amazon Resource Name (ARN) of the quota.

caseId :: Maybe Text

The case ID.

serviceName :: Maybe Text

The service name.

id :: Maybe Text

The unique identifier.

serviceCode :: Maybe Text

The service identifier.

quotaCode :: Maybe Text

The quota identifier.

unit :: Maybe Text

The unit of measurement.

requester :: Maybe Text

The IAM identity of the requester.

quotaName :: Maybe Text

The quota name.

newRequestedServiceQuotaChange :: RequestedServiceQuotaChange

Create a value of RequestedServiceQuotaChange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RequestedServiceQuotaChange', requestedServiceQuotaChange_status - The state of the quota increase request.

$sel:lastUpdated:RequestedServiceQuotaChange', requestedServiceQuotaChange_lastUpdated - The date and time of the most recent change.

$sel:globalQuota:RequestedServiceQuotaChange', requestedServiceQuotaChange_globalQuota - Indicates whether the quota is global.

$sel:created:RequestedServiceQuotaChange', requestedServiceQuotaChange_created - The date and time when the quota increase request was received and the case ID was created.

$sel:desiredValue:RequestedServiceQuotaChange', requestedServiceQuotaChange_desiredValue - The new, increased value for the quota.

$sel:quotaArn:RequestedServiceQuotaChange', requestedServiceQuotaChange_quotaArn - The Amazon Resource Name (ARN) of the quota.

$sel:caseId:RequestedServiceQuotaChange', requestedServiceQuotaChange_caseId - The case ID.

$sel:serviceName:RequestedServiceQuotaChange', requestedServiceQuotaChange_serviceName - The service name.

$sel:id:RequestedServiceQuotaChange', requestedServiceQuotaChange_id - The unique identifier.

$sel:serviceCode:RequestedServiceQuotaChange', requestedServiceQuotaChange_serviceCode - The service identifier.

$sel:quotaCode:RequestedServiceQuotaChange', requestedServiceQuotaChange_quotaCode - The quota identifier.

$sel:unit:RequestedServiceQuotaChange', requestedServiceQuotaChange_unit - The unit of measurement.

$sel:requester:RequestedServiceQuotaChange', requestedServiceQuotaChange_requester - The IAM identity of the requester.

$sel:quotaName:RequestedServiceQuotaChange', requestedServiceQuotaChange_quotaName - The quota name.

requestedServiceQuotaChange_status :: Lens' RequestedServiceQuotaChange (Maybe RequestStatus)

The state of the quota increase request.

requestedServiceQuotaChange_lastUpdated :: Lens' RequestedServiceQuotaChange (Maybe UTCTime)

The date and time of the most recent change.

requestedServiceQuotaChange_globalQuota :: Lens' RequestedServiceQuotaChange (Maybe Bool)

Indicates whether the quota is global.

requestedServiceQuotaChange_created :: Lens' RequestedServiceQuotaChange (Maybe UTCTime)

The date and time when the quota increase request was received and the case ID was created.

requestedServiceQuotaChange_desiredValue :: Lens' RequestedServiceQuotaChange (Maybe Double)

The new, increased value for the quota.

requestedServiceQuotaChange_quotaArn :: Lens' RequestedServiceQuotaChange (Maybe Text)

The Amazon Resource Name (ARN) of the quota.

requestedServiceQuotaChange_caseId :: Lens' RequestedServiceQuotaChange (Maybe Text)

The case ID.

requestedServiceQuotaChange_serviceName :: Lens' RequestedServiceQuotaChange (Maybe Text)

The service name.

requestedServiceQuotaChange_id :: Lens' RequestedServiceQuotaChange (Maybe Text)

The unique identifier.

requestedServiceQuotaChange_serviceCode :: Lens' RequestedServiceQuotaChange (Maybe Text)

The service identifier.

requestedServiceQuotaChange_quotaCode :: Lens' RequestedServiceQuotaChange (Maybe Text)

The quota identifier.

requestedServiceQuotaChange_unit :: Lens' RequestedServiceQuotaChange (Maybe Text)

The unit of measurement.

requestedServiceQuotaChange_requester :: Lens' RequestedServiceQuotaChange (Maybe Text)

The IAM identity of the requester.

requestedServiceQuotaChange_quotaName :: Lens' RequestedServiceQuotaChange (Maybe Text)

The quota name.

ServiceInfo

data ServiceInfo

Information about a service.

See: newServiceInfo smart constructor.

Constructors

ServiceInfo'

Fields

serviceName :: Maybe Text

The service name.

serviceCode :: Maybe Text

The service identifier.

newServiceInfo :: ServiceInfo

Create a value of ServiceInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceName:ServiceInfo', serviceInfo_serviceName - The service name.

$sel:serviceCode:ServiceInfo', serviceInfo_serviceCode - The service identifier.

serviceInfo_serviceName :: Lens' ServiceInfo (Maybe Text)

The service name.

serviceInfo_serviceCode :: Lens' ServiceInfo (Maybe Text)

The service identifier.

ServiceQuota

data ServiceQuota

Information about a quota.

See: newServiceQuota smart constructor.

Constructors

ServiceQuota'

Fields

globalQuota :: Maybe Bool

Indicates whether the quota is global.

period :: Maybe QuotaPeriod

The period of time.

value :: Maybe Double

The quota value.

quotaArn :: Maybe Text

The Amazon Resource Name (ARN) of the quota.

usageMetric :: Maybe MetricInfo

Information about the measurement.

errorReason :: Maybe ErrorReason

The error code and error reason.

adjustable :: Maybe Bool

Indicates whether the quota value can be increased.

serviceName :: Maybe Text

The service name.

serviceCode :: Maybe Text

The service identifier.

quotaCode :: Maybe Text

The quota identifier.

unit :: Maybe Text

The unit of measurement.

quotaName :: Maybe Text

The quota name.

newServiceQuota :: ServiceQuota

Create a value of ServiceQuota with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalQuota:ServiceQuota', serviceQuota_globalQuota - Indicates whether the quota is global.

$sel:period:ServiceQuota', serviceQuota_period - The period of time.

$sel:value:ServiceQuota', serviceQuota_value - The quota value.

$sel:quotaArn:ServiceQuota', serviceQuota_quotaArn - The Amazon Resource Name (ARN) of the quota.

$sel:usageMetric:ServiceQuota', serviceQuota_usageMetric - Information about the measurement.

$sel:errorReason:ServiceQuota', serviceQuota_errorReason - The error code and error reason.

$sel:adjustable:ServiceQuota', serviceQuota_adjustable - Indicates whether the quota value can be increased.

$sel:serviceName:ServiceQuota', serviceQuota_serviceName - The service name.

$sel:serviceCode:ServiceQuota', serviceQuota_serviceCode - The service identifier.

$sel:quotaCode:ServiceQuota', serviceQuota_quotaCode - The quota identifier.

$sel:unit:ServiceQuota', serviceQuota_unit - The unit of measurement.

$sel:quotaName:ServiceQuota', serviceQuota_quotaName - The quota name.

serviceQuota_globalQuota :: Lens' ServiceQuota (Maybe Bool)

Indicates whether the quota is global.

serviceQuota_period :: Lens' ServiceQuota (Maybe QuotaPeriod)

The period of time.

serviceQuota_value :: Lens' ServiceQuota (Maybe Double)

The quota value.

serviceQuota_quotaArn :: Lens' ServiceQuota (Maybe Text)

The Amazon Resource Name (ARN) of the quota.

serviceQuota_usageMetric :: Lens' ServiceQuota (Maybe MetricInfo)

Information about the measurement.

serviceQuota_errorReason :: Lens' ServiceQuota (Maybe ErrorReason)

The error code and error reason.

serviceQuota_adjustable :: Lens' ServiceQuota (Maybe Bool)

Indicates whether the quota value can be increased.

serviceQuota_serviceName :: Lens' ServiceQuota (Maybe Text)

The service name.

serviceQuota_serviceCode :: Lens' ServiceQuota (Maybe Text)

The service identifier.

serviceQuota_quotaCode :: Lens' ServiceQuota (Maybe Text)

The quota identifier.

serviceQuota_unit :: Lens' ServiceQuota (Maybe Text)

The unit of measurement.

serviceQuota_quotaName :: Lens' ServiceQuota (Maybe Text)

The quota name.

ServiceQuotaIncreaseRequestInTemplate

data ServiceQuotaIncreaseRequestInTemplate

Information about a quota increase request.

See: newServiceQuotaIncreaseRequestInTemplate smart constructor.

Constructors

ServiceQuotaIncreaseRequestInTemplate'

Fields

globalQuota :: Maybe Bool

Indicates whether the quota is global.

desiredValue :: Maybe Double

The new, increased value of the quota.

serviceName :: Maybe Text

The service name.

awsRegion :: Maybe Text

The AWS Region.

serviceCode :: Maybe Text

The service identifier.

quotaCode :: Maybe Text

The quota identifier.

unit :: Maybe Text

The unit of measurement.

quotaName :: Maybe Text

The quota name.

newServiceQuotaIncreaseRequestInTemplate :: ServiceQuotaIncreaseRequestInTemplate

Create a value of ServiceQuotaIncreaseRequestInTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalQuota:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_globalQuota - Indicates whether the quota is global.

$sel:desiredValue:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_desiredValue - The new, increased value of the quota.

$sel:serviceName:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_serviceName - The service name.

$sel:awsRegion:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_awsRegion - The AWS Region.

$sel:serviceCode:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_serviceCode - The service identifier.

$sel:quotaCode:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_quotaCode - The quota identifier.

$sel:unit:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_unit - The unit of measurement.

$sel:quotaName:ServiceQuotaIncreaseRequestInTemplate', serviceQuotaIncreaseRequestInTemplate_quotaName - The quota name.

serviceQuotaIncreaseRequestInTemplate_globalQuota :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Bool)

Indicates whether the quota is global.

serviceQuotaIncreaseRequestInTemplate_desiredValue :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Double)

The new, increased value of the quota.

serviceQuotaIncreaseRequestInTemplate_serviceName :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Text)

The service name.

serviceQuotaIncreaseRequestInTemplate_awsRegion :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Text)

The AWS Region.

serviceQuotaIncreaseRequestInTemplate_serviceCode :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Text)

The service identifier.

serviceQuotaIncreaseRequestInTemplate_quotaCode :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Text)

The quota identifier.

serviceQuotaIncreaseRequestInTemplate_unit :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Text)

The unit of measurement.

serviceQuotaIncreaseRequestInTemplate_quotaName :: Lens' ServiceQuotaIncreaseRequestInTemplate (Maybe Text)

The quota name.

Tag

data Tag

A complex data type that contains a tag key and tag value.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

A string that contains a tag key. The string length should be between 1 and 128 characters. Valid characters include a-z, A-Z, 0-9, space, and the special characters _ - . : / = + @.

value :: Text

A string that contains an optional tag value. The string length should be between 0 and 256 characters. Valid characters include a-z, A-Z, 0-9, space, and the special characters _ - . : / = + @.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - A string that contains a tag key. The string length should be between 1 and 128 characters. Valid characters include a-z, A-Z, 0-9, space, and the special characters _ - . : / = + @.

$sel:value:Tag', tag_value - A string that contains an optional tag value. The string length should be between 0 and 256 characters. Valid characters include a-z, A-Z, 0-9, space, and the special characters _ - . : / = + @.

tag_key :: Lens' Tag Text

A string that contains a tag key. The string length should be between 1 and 128 characters. Valid characters include a-z, A-Z, 0-9, space, and the special characters _ - . : / = + @.

tag_value :: Lens' Tag Text

A string that contains an optional tag value. The string length should be between 0 and 256 characters. Valid characters include a-z, A-Z, 0-9, space, and the special characters _ - . : / = + @.