Amazonka.Shield.Types.AttackProperty

Description

Documentation

data AttackProperty

Details of a Shield event. This is provided as part of an AttackDetail.

See: newAttackProperty smart constructor.

Constructors

AttackProperty'

Fields

attackLayer :: Maybe AttackLayer

The type of Shield event that was observed. NETWORK indicates layer 3 and layer 4 events and APPLICATION indicates layer 7 events.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

topContributors :: Maybe [Contributor]

Contributor objects for the top five contributors to a Shield event.

attackPropertyIdentifier :: Maybe AttackPropertyIdentifier

Defines the Shield event property information that is provided. The WORDPRESS_PINGBACK_REFLECTOR and WORDPRESS_PINGBACK_SOURCE values are valid only for WordPress reflective pingback events.

total :: Maybe Integer

The total contributions made to this Shield event by all contributors.

unit :: Maybe Unit

The unit used for the Contributor Value property.

newAttackProperty :: AttackProperty

Create a value of AttackProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attackLayer:AttackProperty', attackProperty_attackLayer - The type of Shield event that was observed. NETWORK indicates layer 3 and layer 4 events and APPLICATION indicates layer 7 events.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

$sel:topContributors:AttackProperty', attackProperty_topContributors - Contributor objects for the top five contributors to a Shield event.

$sel:attackPropertyIdentifier:AttackProperty', attackProperty_attackPropertyIdentifier - Defines the Shield event property information that is provided. The WORDPRESS_PINGBACK_REFLECTOR and WORDPRESS_PINGBACK_SOURCE values are valid only for WordPress reflective pingback events.

$sel:total:AttackProperty', attackProperty_total - The total contributions made to this Shield event by all contributors.

$sel:unit:AttackProperty', attackProperty_unit - The unit used for the Contributor Value property.

attackProperty_attackLayer :: Lens' AttackProperty (Maybe AttackLayer)

The type of Shield event that was observed. NETWORK indicates layer 3 and layer 4 events and APPLICATION indicates layer 7 events.

For infrastructure layer events (L3 and L4 events) after January 25, 2021, you can view metrics for top contributors in Amazon CloudWatch metrics. For more information, see Shield metrics and alarms in the WAF Developer Guide.

attackProperty_topContributors :: Lens' AttackProperty (Maybe [Contributor])

Contributor objects for the top five contributors to a Shield event.

attackProperty_attackPropertyIdentifier :: Lens' AttackProperty (Maybe AttackPropertyIdentifier)

Defines the Shield event property information that is provided. The WORDPRESS_PINGBACK_REFLECTOR and WORDPRESS_PINGBACK_SOURCE values are valid only for WordPress reflective pingback events.

attackProperty_total :: Lens' AttackProperty (Maybe Integer)

The total contributions made to this Shield event by all contributors.

attackProperty_unit :: Lens' AttackProperty (Maybe Unit)

The unit used for the Contributor Value property.