Amazonka.Shield.Types.AttackVolume

Description

Documentation

data AttackVolume

Information about the volume of attacks during the time period, included in an AttackStatisticsDataItem. If the accompanying AttackCount in the statistics object is zero, this setting might be empty.

See: newAttackVolume smart constructor.

Constructors

AttackVolume'

Fields

packetsPerSecond :: Maybe AttackVolumeStatistics

A statistics object that uses packets per second as the unit. This is included for network level attacks.

requestsPerSecond :: Maybe AttackVolumeStatistics

A statistics object that uses requests per second as the unit. This is included for application level attacks, and is only available for accounts that are subscribed to Shield Advanced.

bitsPerSecond :: Maybe AttackVolumeStatistics

A statistics object that uses bits per second as the unit. This is included for network level attacks.

newAttackVolume :: AttackVolume

Create a value of AttackVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packetsPerSecond:AttackVolume', attackVolume_packetsPerSecond - A statistics object that uses packets per second as the unit. This is included for network level attacks.

$sel:requestsPerSecond:AttackVolume', attackVolume_requestsPerSecond - A statistics object that uses requests per second as the unit. This is included for application level attacks, and is only available for accounts that are subscribed to Shield Advanced.

$sel:bitsPerSecond:AttackVolume', attackVolume_bitsPerSecond - A statistics object that uses bits per second as the unit. This is included for network level attacks.

attackVolume_packetsPerSecond :: Lens' AttackVolume (Maybe AttackVolumeStatistics)

A statistics object that uses packets per second as the unit. This is included for network level attacks.

attackVolume_requestsPerSecond :: Lens' AttackVolume (Maybe AttackVolumeStatistics)

A statistics object that uses requests per second as the unit. This is included for application level attacks, and is only available for accounts that are subscribed to Shield Advanced.

attackVolume_bitsPerSecond :: Lens' AttackVolume (Maybe AttackVolumeStatistics)

A statistics object that uses bits per second as the unit. This is included for network level attacks.