Amazonka.Sign.V2Header

Description

This module provides an AWS compliant V2 Header request signer. It is based heavily on boto, specifically boto's HmacAuthV1Handler AWS capable signer. AWS documentation is available here.

Notice: Limitations include an inability to sign with a security token and inability to overwrite the Date header with an expiry.

Documentation

v2Header :: Signer

newSigner :: RequestHeaders -> ByteString -> ByteString -> QueryString -> ByteString

Construct a full header signer following the V2 Header scheme

toSignerQueryBS :: QueryString -> ByteString

The following function mostly follows the toBS in amazonka QueryString except for single QValue or single QPair keys not being suffixed with an equals.

constructSigningHeader :: Header -> ByteString

Construct a header string for signing

constructSigningQuery :: QueryString -> QueryString

Constructs a query string for signing

constructFullPath :: ByteString -> ByteString -> ByteString

unionNecessaryHeaders :: [Header] -> [Header]