Amazonka.Sign.V4.Base

Description

Documentation

data V4

Constructors

V4

Fields

metaTime :: UTCTime

metaMethod :: Method

metaPath :: Path

metaEndpoint :: Endpoint

metaCredential :: Credential

metaCanonicalQuery :: CanonicalQuery

metaCanonicalRequest :: CanonicalRequest

metaCanonicalHeaders :: CanonicalHeaders

metaSignedHeaders :: SignedHeaders

metaStringToSign :: StringToSign

metaSignature :: Signature

metaHeaders :: [Header]

metaTimeout :: Maybe Seconds

base :: Hash -> Request a -> AuthEnv -> Region -> UTCTime -> (V4, ClientRequest -> ClientRequest)

newtype Tag (s :: Symbol) a

Used to tag provenance. This allows keeping the same layout as the signing documentation, passing ByteStrings everywhere, with some type guarantees.

Data.Tagged is not used for no reason other than the dependency, syntactic length, and the ToByteString instance.

Constructors

Tag

Fields

untag :: a

type Hash = Tag "body-digest" ByteString

type StringToSign = Tag "string-to-sign" ByteString

type Credential = Tag "credential" ByteString

type CredentialScope = Tag "credential-scope" [ByteString]

type CanonicalRequest = Tag "canonical-request" ByteString

type CanonicalHeaders = Tag "canonical-headers" ByteString

type CanonicalQuery = Tag "canonical-query" ByteString

type SignedHeaders = Tag "signed-headers" ByteString

type NormalisedHeaders = Tag "normalised-headers" [(ByteString, ByteString)]

type Method = Tag "method" ByteString

type Path = Tag "path" ByteString

type Signature = Tag "signature" ByteString

authorisation :: V4 -> ByteString

signRequest

Arguments

:: V4
Pre-signRequestd signing metadata.

-> RequestBody
The request body.

-> (ClientRequest -> ClientRequest)
Insert authentication information.

-> Signed a

signMetadata :: AuthEnv -> Region -> UTCTime -> (Credential -> SignedHeaders -> QueryString -> QueryString) -> Hash -> Request a -> V4

algorithm :: ByteString

signature :: SecretKey -> CredentialScope -> StringToSign -> Signature

stringToSign :: UTCTime -> CredentialScope -> CanonicalRequest -> StringToSign

credential :: AccessKey -> CredentialScope -> Credential

credentialScope :: Service -> Endpoint -> UTCTime -> CredentialScope

canonicalRequest :: Method -> Path -> Hash -> CanonicalQuery -> CanonicalHeaders -> SignedHeaders -> CanonicalRequest

escapedPath :: Request a -> Path

canonicalQuery :: QueryString -> CanonicalQuery

canonicalHeaders :: NormalisedHeaders -> CanonicalHeaders

signedHeaders :: NormalisedHeaders -> SignedHeaders

normaliseHeaders :: [Header] -> NormalisedHeaders