Amazonka.Signer.AddProfilePermission

Description

Adds cross-account permissions to a signing profile.

Creating a Request

data AddProfilePermission

See: newAddProfilePermission smart constructor.

Constructors

AddProfilePermission'

Fields

profileVersion :: Maybe Text

The version of the signing profile.

revisionId :: Maybe Text

A unique identifier for the current profile revision.

action :: Text

The AWS Signer action permitted as part of cross-account permissions.

principal :: Text

The AWS principal receiving cross-account permissions. This may be an IAM role or another AWS account ID.

statementId :: Text

A unique identifier for the cross-account permission statement.

profileName :: Text

The human-readable name of the signing profile.

newAddProfilePermission

Arguments

:: Text
$sel:action:AddProfilePermission'

-> Text
$sel:principal:AddProfilePermission'

-> Text
$sel:statementId:AddProfilePermission'

-> Text
$sel:profileName:AddProfilePermission'

-> AddProfilePermission

Create a value of AddProfilePermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profileVersion:AddProfilePermission', addProfilePermission_profileVersion - The version of the signing profile.

$sel:revisionId:AddProfilePermission', addProfilePermission_revisionId - A unique identifier for the current profile revision.

$sel:action:AddProfilePermission', addProfilePermission_action - The AWS Signer action permitted as part of cross-account permissions.

$sel:principal:AddProfilePermission', addProfilePermission_principal - The AWS principal receiving cross-account permissions. This may be an IAM role or another AWS account ID.

$sel:statementId:AddProfilePermission', addProfilePermission_statementId - A unique identifier for the cross-account permission statement.

$sel:profileName:AddProfilePermission', addProfilePermission_profileName - The human-readable name of the signing profile.

Request Lenses

addProfilePermission_profileVersion :: Lens' AddProfilePermission (Maybe Text)

The version of the signing profile.

addProfilePermission_revisionId :: Lens' AddProfilePermission (Maybe Text)

A unique identifier for the current profile revision.

addProfilePermission_action :: Lens' AddProfilePermission Text

The AWS Signer action permitted as part of cross-account permissions.

addProfilePermission_principal :: Lens' AddProfilePermission Text

The AWS principal receiving cross-account permissions. This may be an IAM role or another AWS account ID.

addProfilePermission_statementId :: Lens' AddProfilePermission Text

A unique identifier for the cross-account permission statement.

addProfilePermission_profileName :: Lens' AddProfilePermission Text

The human-readable name of the signing profile.

Destructuring the Response

data AddProfilePermissionResponse

See: newAddProfilePermissionResponse smart constructor.

Constructors

AddProfilePermissionResponse'

Fields

revisionId :: Maybe Text

A unique identifier for the current profile revision.

httpStatus :: Int

The response's http status code.

newAddProfilePermissionResponse

Arguments

:: Int
$sel:httpStatus:AddProfilePermissionResponse'

-> AddProfilePermissionResponse

Create a value of AddProfilePermissionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revisionId:AddProfilePermission', addProfilePermissionResponse_revisionId - A unique identifier for the current profile revision.

$sel:httpStatus:AddProfilePermissionResponse', addProfilePermissionResponse_httpStatus - The response's http status code.

Response Lenses

addProfilePermissionResponse_revisionId :: Lens' AddProfilePermissionResponse (Maybe Text)

A unique identifier for the current profile revision.

addProfilePermissionResponse_httpStatus :: Lens' AddProfilePermissionResponse Int

The response's http status code.