Amazonka.Signer.ListSigningProfiles

Description

Lists all available signing profiles in your AWS account. Returns only profiles with an ACTIVE status unless the includeCanceled request field is set to true. If additional jobs remain to be listed, code signing returns a nextToken value. Use this value in subsequent calls to ListSigningJobs to fetch the remaining values. You can continue calling ListSigningJobs with your maxResults parameter and with new values that code signing returns in the nextToken parameter until all of your signing jobs have been returned.

This operation returns paginated results.

Creating a Request

data ListSigningProfiles

See: newListSigningProfiles smart constructor.

Constructors

ListSigningProfiles'

Fields

nextToken :: Maybe Text

Value for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of nextToken from the response that you just received.

platformId :: Maybe Text

Filters results to return only signing jobs initiated for a specified signing platform.

statuses :: Maybe [SigningProfileStatus]

Filters results to return only signing jobs with statuses in the specified list.

includeCanceled :: Maybe Bool

Designates whether to include profiles with the status of CANCELED.

maxResults :: Maybe Natural

The maximum number of profiles to be returned.

newListSigningProfiles :: ListSigningProfiles

Create a value of ListSigningProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSigningProfiles', listSigningProfiles_nextToken - Value for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of nextToken from the response that you just received.

$sel:platformId:ListSigningProfiles', listSigningProfiles_platformId - Filters results to return only signing jobs initiated for a specified signing platform.

$sel:statuses:ListSigningProfiles', listSigningProfiles_statuses - Filters results to return only signing jobs with statuses in the specified list.

$sel:includeCanceled:ListSigningProfiles', listSigningProfiles_includeCanceled - Designates whether to include profiles with the status of CANCELED.

$sel:maxResults:ListSigningProfiles', listSigningProfiles_maxResults - The maximum number of profiles to be returned.

Request Lenses

listSigningProfiles_nextToken :: Lens' ListSigningProfiles (Maybe Text)

Value for specifying the next set of paginated results to return. After you receive a response with truncated results, use this parameter in a subsequent request. Set it to the value of nextToken from the response that you just received.

listSigningProfiles_platformId :: Lens' ListSigningProfiles (Maybe Text)

Filters results to return only signing jobs initiated for a specified signing platform.

listSigningProfiles_statuses :: Lens' ListSigningProfiles (Maybe [SigningProfileStatus])

Filters results to return only signing jobs with statuses in the specified list.

listSigningProfiles_includeCanceled :: Lens' ListSigningProfiles (Maybe Bool)

Designates whether to include profiles with the status of CANCELED.

listSigningProfiles_maxResults :: Lens' ListSigningProfiles (Maybe Natural)

The maximum number of profiles to be returned.

Destructuring the Response

data ListSigningProfilesResponse

See: newListSigningProfilesResponse smart constructor.

Constructors

ListSigningProfilesResponse'

Fields

profiles :: Maybe [SigningProfile]

A list of profiles that are available in the AWS account. This includes profiles with the status of CANCELED if the includeCanceled parameter is set to true.

nextToken :: Maybe Text

Value for specifying the next set of paginated results to return.

httpStatus :: Int

The response's http status code.

newListSigningProfilesResponse

Arguments

:: Int
$sel:httpStatus:ListSigningProfilesResponse'

-> ListSigningProfilesResponse

Create a value of ListSigningProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profiles:ListSigningProfilesResponse', listSigningProfilesResponse_profiles - A list of profiles that are available in the AWS account. This includes profiles with the status of CANCELED if the includeCanceled parameter is set to true.

$sel:nextToken:ListSigningProfiles', listSigningProfilesResponse_nextToken - Value for specifying the next set of paginated results to return.

$sel:httpStatus:ListSigningProfilesResponse', listSigningProfilesResponse_httpStatus - The response's http status code.

Response Lenses

listSigningProfilesResponse_profiles :: Lens' ListSigningProfilesResponse (Maybe [SigningProfile])

A list of profiles that are available in the AWS account. This includes profiles with the status of CANCELED if the includeCanceled parameter is set to true.

listSigningProfilesResponse_nextToken :: Lens' ListSigningProfilesResponse (Maybe Text)

Value for specifying the next set of paginated results to return.

listSigningProfilesResponse_httpStatus :: Lens' ListSigningProfilesResponse Int

The response's http status code.