Amazonka.Signer.StartSigningJob

Description

Initiates a signing job to be performed on the code provided. Signing jobs are viewable by the ListSigningJobs operation for two years after they are performed. Note the following requirements:

You must create an Amazon S3 source bucket. For more information, see Create a Bucket in the Amazon S3 Getting Started Guide.

Your S3 source bucket must be version enabled.

You must create an S3 destination bucket. Code signing uses your S3 destination bucket to write your signed code.

You specify the name of the source and destination buckets when calling the StartSigningJob operation.

You must also specify a request token that identifies your request to code signing.

You can call the DescribeSigningJob and the ListSigningJobs actions after you call StartSigningJob.

For a Java example that shows how to use this action, see http://docs.aws.amazon.com/acm/latest/userguide/

Creating a Request

data StartSigningJob

See: newStartSigningJob smart constructor.

Constructors

StartSigningJob'

Fields

profileOwner :: Maybe Text

The AWS account ID of the signing profile owner.

source :: Source

The S3 bucket that contains the object to sign or a BLOB that contains your raw code.

destination :: Destination

The S3 bucket in which to save your signed object. The destination contains the name of your bucket and an optional prefix.

profileName :: Text

The name of the signing profile.

clientRequestToken :: Text

String that identifies the signing request. All calls after the first that use this token return the same response as the first call.

newStartSigningJob

Arguments

:: Source
$sel:source:StartSigningJob'

-> Destination
$sel:destination:StartSigningJob'

-> Text
$sel:profileName:StartSigningJob'

-> Text
$sel:clientRequestToken:StartSigningJob'

-> StartSigningJob

Create a value of StartSigningJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profileOwner:StartSigningJob', startSigningJob_profileOwner - The AWS account ID of the signing profile owner.

$sel:source:StartSigningJob', startSigningJob_source - The S3 bucket that contains the object to sign or a BLOB that contains your raw code.

$sel:destination:StartSigningJob', startSigningJob_destination - The S3 bucket in which to save your signed object. The destination contains the name of your bucket and an optional prefix.

$sel:profileName:StartSigningJob', startSigningJob_profileName - The name of the signing profile.

$sel:clientRequestToken:StartSigningJob', startSigningJob_clientRequestToken - String that identifies the signing request. All calls after the first that use this token return the same response as the first call.

Request Lenses

startSigningJob_profileOwner :: Lens' StartSigningJob (Maybe Text)

The AWS account ID of the signing profile owner.

startSigningJob_source :: Lens' StartSigningJob Source

The S3 bucket that contains the object to sign or a BLOB that contains your raw code.

startSigningJob_destination :: Lens' StartSigningJob Destination

The S3 bucket in which to save your signed object. The destination contains the name of your bucket and an optional prefix.

startSigningJob_profileName :: Lens' StartSigningJob Text

The name of the signing profile.

startSigningJob_clientRequestToken :: Lens' StartSigningJob Text

String that identifies the signing request. All calls after the first that use this token return the same response as the first call.

Destructuring the Response

data StartSigningJobResponse

See: newStartSigningJobResponse smart constructor.

Constructors

StartSigningJobResponse'

Fields

jobId :: Maybe Text

The ID of your signing job.

jobOwner :: Maybe Text

The AWS account ID of the signing job owner.

httpStatus :: Int

The response's http status code.

newStartSigningJobResponse

Arguments

:: Int
$sel:httpStatus:StartSigningJobResponse'

-> StartSigningJobResponse

Create a value of StartSigningJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StartSigningJobResponse', startSigningJobResponse_jobId - The ID of your signing job.

$sel:jobOwner:StartSigningJobResponse', startSigningJobResponse_jobOwner - The AWS account ID of the signing job owner.

$sel:httpStatus:StartSigningJobResponse', startSigningJobResponse_httpStatus - The response's http status code.

Response Lenses

startSigningJobResponse_jobId :: Lens' StartSigningJobResponse (Maybe Text)

The ID of your signing job.

startSigningJobResponse_jobOwner :: Lens' StartSigningJobResponse (Maybe Text)

The AWS account ID of the signing job owner.

startSigningJobResponse_httpStatus :: Lens' StartSigningJobResponse Int

The response's http status code.