Amazonka.Signer.Types

Description

Service Configuration

defaultService :: Service

API version 2017-08-25 of the Amazon Signer SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

You signing certificate could not be validated.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The resource encountered a conflicting state.

_ServiceLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The client is making a request that exceeds service limits.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The signing profile was not found.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

Instead of this error, TooManyRequestsException should be used.

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

The allowed number of job-signing requests has been exceeded.

This error supersedes the error ThrottlingException.

_InternalServiceErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error occurred.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

A specified resource could not be found.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request contains invalid parameters for the ARN or tags. This exception also occurs when you call a tagging API on a cancelled signing profile.

Category

newtype Category

Constructors

Category'

Fields

fromCategory :: Text

Bundled Patterns

pattern Category_AWSIoT :: Category

EncryptionAlgorithm

newtype EncryptionAlgorithm

Constructors

EncryptionAlgorithm'

Fields

fromEncryptionAlgorithm :: Text

Bundled Patterns

pattern EncryptionAlgorithm_ECDSA :: EncryptionAlgorithm

pattern EncryptionAlgorithm_RSA :: EncryptionAlgorithm

HashAlgorithm

newtype HashAlgorithm

Constructors

HashAlgorithm'

Fields

fromHashAlgorithm :: Text

Bundled Patterns

pattern HashAlgorithm_SHA1 :: HashAlgorithm

pattern HashAlgorithm_SHA256 :: HashAlgorithm

ImageFormat

newtype ImageFormat

Constructors

ImageFormat'

Fields

fromImageFormat :: Text

Bundled Patterns

pattern ImageFormat_JSON :: ImageFormat

pattern ImageFormat_JSONDetached :: ImageFormat

pattern ImageFormat_JSONEmbedded :: ImageFormat

SigningProfileStatus

newtype SigningProfileStatus

Constructors

SigningProfileStatus'

Fields

fromSigningProfileStatus :: Text

Bundled Patterns

pattern SigningProfileStatus_Active :: SigningProfileStatus

pattern SigningProfileStatus_Canceled :: SigningProfileStatus

pattern SigningProfileStatus_Revoked :: SigningProfileStatus

SigningStatus

newtype SigningStatus

Constructors

SigningStatus'

Fields

fromSigningStatus :: Text

Bundled Patterns

pattern SigningStatus_Failed :: SigningStatus

pattern SigningStatus_InProgress :: SigningStatus

pattern SigningStatus_Succeeded :: SigningStatus

ValidityType

newtype ValidityType

Constructors

ValidityType'

Fields

fromValidityType :: Text

Bundled Patterns

pattern ValidityType_DAYS :: ValidityType

pattern ValidityType_MONTHS :: ValidityType

pattern ValidityType_YEARS :: ValidityType

Destination

data Destination

Points to an S3Destination object that contains information about your S3 bucket.

See: newDestination smart constructor.

Constructors

Destination'

Fields

s3 :: Maybe S3Destination

The S3Destination object.

newDestination :: Destination

Create a value of Destination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3:Destination', destination_s3 - The S3Destination object.

destination_s3 :: Lens' Destination (Maybe S3Destination)

The S3Destination object.

EncryptionAlgorithmOptions

data EncryptionAlgorithmOptions

The encryption algorithm options that are available to a code signing job.

See: newEncryptionAlgorithmOptions smart constructor.

Constructors

EncryptionAlgorithmOptions'

Fields

allowedValues :: [EncryptionAlgorithm]

The set of accepted encryption algorithms that are allowed in a code signing job.

defaultValue :: EncryptionAlgorithm

The default encryption algorithm that is used by a code signing job.

newEncryptionAlgorithmOptions

Arguments

:: EncryptionAlgorithm
$sel:defaultValue:EncryptionAlgorithmOptions'

-> EncryptionAlgorithmOptions

Create a value of EncryptionAlgorithmOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowedValues:EncryptionAlgorithmOptions', encryptionAlgorithmOptions_allowedValues - The set of accepted encryption algorithms that are allowed in a code signing job.

$sel:defaultValue:EncryptionAlgorithmOptions', encryptionAlgorithmOptions_defaultValue - The default encryption algorithm that is used by a code signing job.

encryptionAlgorithmOptions_allowedValues :: Lens' EncryptionAlgorithmOptions [EncryptionAlgorithm]

The set of accepted encryption algorithms that are allowed in a code signing job.

encryptionAlgorithmOptions_defaultValue :: Lens' EncryptionAlgorithmOptions EncryptionAlgorithm

The default encryption algorithm that is used by a code signing job.

HashAlgorithmOptions

data HashAlgorithmOptions

The hash algorithms that are available to a code signing job.

See: newHashAlgorithmOptions smart constructor.

Constructors

HashAlgorithmOptions'

Fields

allowedValues :: [HashAlgorithm]

The set of accepted hash algorithms allowed in a code signing job.

defaultValue :: HashAlgorithm

The default hash algorithm that is used in a code signing job.

newHashAlgorithmOptions

Arguments

:: HashAlgorithm
$sel:defaultValue:HashAlgorithmOptions'

-> HashAlgorithmOptions

Create a value of HashAlgorithmOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowedValues:HashAlgorithmOptions', hashAlgorithmOptions_allowedValues - The set of accepted hash algorithms allowed in a code signing job.

$sel:defaultValue:HashAlgorithmOptions', hashAlgorithmOptions_defaultValue - The default hash algorithm that is used in a code signing job.

hashAlgorithmOptions_allowedValues :: Lens' HashAlgorithmOptions [HashAlgorithm]

The set of accepted hash algorithms allowed in a code signing job.

hashAlgorithmOptions_defaultValue :: Lens' HashAlgorithmOptions HashAlgorithm

The default hash algorithm that is used in a code signing job.

Permission

data Permission

A cross-account permission for a signing profile.

See: newPermission smart constructor.

Constructors

Permission'

Fields

action :: Maybe Text

An AWS Signer action permitted as part of cross-account permissions.

profileVersion :: Maybe Text

The signing profile version that a permission applies to.

principal :: Maybe Text

The AWS principal that has been granted a cross-account permission.

statementId :: Maybe Text

A unique identifier for a cross-account permission statement.

newPermission :: Permission

Create a value of Permission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:Permission', permission_action - An AWS Signer action permitted as part of cross-account permissions.

$sel:profileVersion:Permission', permission_profileVersion - The signing profile version that a permission applies to.

$sel:principal:Permission', permission_principal - The AWS principal that has been granted a cross-account permission.

$sel:statementId:Permission', permission_statementId - A unique identifier for a cross-account permission statement.

permission_action :: Lens' Permission (Maybe Text)

An AWS Signer action permitted as part of cross-account permissions.

permission_profileVersion :: Lens' Permission (Maybe Text)

The signing profile version that a permission applies to.

permission_principal :: Lens' Permission (Maybe Text)

The AWS principal that has been granted a cross-account permission.

permission_statementId :: Lens' Permission (Maybe Text)

A unique identifier for a cross-account permission statement.

S3Destination

data S3Destination

The name and prefix of the S3 bucket where code signing saves your signed objects.

See: newS3Destination smart constructor.

Constructors

S3Destination'

Fields

prefix :: Maybe Text

An Amazon S3 prefix that you can use to limit responses to those that begin with the specified prefix.

bucketName :: Maybe Text

Name of the S3 bucket.

newS3Destination :: S3Destination

Create a value of S3Destination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:S3Destination', s3Destination_prefix - An Amazon S3 prefix that you can use to limit responses to those that begin with the specified prefix.

$sel:bucketName:S3Destination', s3Destination_bucketName - Name of the S3 bucket.

s3Destination_prefix :: Lens' S3Destination (Maybe Text)

An Amazon S3 prefix that you can use to limit responses to those that begin with the specified prefix.

s3Destination_bucketName :: Lens' S3Destination (Maybe Text)

Name of the S3 bucket.

S3SignedObject

data S3SignedObject

The S3 bucket name and key where code signing saved your signed code image.

See: newS3SignedObject smart constructor.

Constructors

S3SignedObject'

Fields

bucketName :: Maybe Text

Name of the S3 bucket.

key :: Maybe Text

Key name that uniquely identifies a signed code image in your bucket.

newS3SignedObject :: S3SignedObject

Create a value of S3SignedObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketName:S3SignedObject', s3SignedObject_bucketName - Name of the S3 bucket.

$sel:key:S3SignedObject', s3SignedObject_key - Key name that uniquely identifies a signed code image in your bucket.

s3SignedObject_bucketName :: Lens' S3SignedObject (Maybe Text)

Name of the S3 bucket.

s3SignedObject_key :: Lens' S3SignedObject (Maybe Text)

Key name that uniquely identifies a signed code image in your bucket.

S3Source

data S3Source

Information about the S3 bucket where you saved your unsigned code.

See: newS3Source smart constructor.

Constructors

S3Source'

Fields

bucketName :: Text

Name of the S3 bucket.

key :: Text

Key name of the bucket object that contains your unsigned code.

version :: Text

Version of your source image in your version enabled S3 bucket.

newS3Source

Arguments

:: Text
$sel:bucketName:S3Source'

-> Text
$sel:key:S3Source'

-> Text
$sel:version:S3Source'

-> S3Source

Create a value of S3Source with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketName:S3Source', s3Source_bucketName - Name of the S3 bucket.

$sel:key:S3Source', s3Source_key - Key name of the bucket object that contains your unsigned code.

$sel:version:S3Source', s3Source_version - Version of your source image in your version enabled S3 bucket.

s3Source_bucketName :: Lens' S3Source Text

Name of the S3 bucket.

s3Source_key :: Lens' S3Source Text

Key name of the bucket object that contains your unsigned code.

s3Source_version :: Lens' S3Source Text

Version of your source image in your version enabled S3 bucket.

SignatureValidityPeriod

data SignatureValidityPeriod

The validity period for a signing job.

See: newSignatureValidityPeriod smart constructor.

Constructors

SignatureValidityPeriod'

Fields

value :: Maybe Int

The numerical value of the time unit for signature validity.

type' :: Maybe ValidityType

The time unit for signature validity.

newSignatureValidityPeriod :: SignatureValidityPeriod

Create a value of SignatureValidityPeriod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:SignatureValidityPeriod', signatureValidityPeriod_value - The numerical value of the time unit for signature validity.

$sel:type':SignatureValidityPeriod', signatureValidityPeriod_type - The time unit for signature validity.

signatureValidityPeriod_value :: Lens' SignatureValidityPeriod (Maybe Int)

The numerical value of the time unit for signature validity.

signatureValidityPeriod_type :: Lens' SignatureValidityPeriod (Maybe ValidityType)

The time unit for signature validity.

SignedObject

data SignedObject

Points to an S3SignedObject object that contains information about your signed code image.

See: newSignedObject smart constructor.

Constructors

SignedObject'

Fields

s3 :: Maybe S3SignedObject

The S3SignedObject.

newSignedObject :: SignedObject

Create a value of SignedObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3:SignedObject', signedObject_s3 - The S3SignedObject.

signedObject_s3 :: Lens' SignedObject (Maybe S3SignedObject)

The S3SignedObject.

SigningConfiguration

data SigningConfiguration

The configuration of a code signing operation.

See: newSigningConfiguration smart constructor.

Constructors

SigningConfiguration'

Fields

encryptionAlgorithmOptions :: EncryptionAlgorithmOptions

The encryption algorithm options that are available for a code signing job.

hashAlgorithmOptions :: HashAlgorithmOptions

The hash algorithm options that are available for a code signing job.

newSigningConfiguration

Arguments

:: EncryptionAlgorithmOptions
$sel:encryptionAlgorithmOptions:SigningConfiguration'

-> HashAlgorithmOptions
$sel:hashAlgorithmOptions:SigningConfiguration'

-> SigningConfiguration

Create a value of SigningConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionAlgorithmOptions:SigningConfiguration', signingConfiguration_encryptionAlgorithmOptions - The encryption algorithm options that are available for a code signing job.

$sel:hashAlgorithmOptions:SigningConfiguration', signingConfiguration_hashAlgorithmOptions - The hash algorithm options that are available for a code signing job.

signingConfiguration_encryptionAlgorithmOptions :: Lens' SigningConfiguration EncryptionAlgorithmOptions

The encryption algorithm options that are available for a code signing job.

signingConfiguration_hashAlgorithmOptions :: Lens' SigningConfiguration HashAlgorithmOptions

The hash algorithm options that are available for a code signing job.

SigningConfigurationOverrides

data SigningConfigurationOverrides

A signing configuration that overrides the default encryption or hash algorithm of a signing job.

See: newSigningConfigurationOverrides smart constructor.

Constructors

SigningConfigurationOverrides'

Fields

hashAlgorithm :: Maybe HashAlgorithm

A specified override of the default hash algorithm that is used in a code signing job.

encryptionAlgorithm :: Maybe EncryptionAlgorithm

A specified override of the default encryption algorithm that is used in a code signing job.

newSigningConfigurationOverrides :: SigningConfigurationOverrides

Create a value of SigningConfigurationOverrides with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hashAlgorithm:SigningConfigurationOverrides', signingConfigurationOverrides_hashAlgorithm - A specified override of the default hash algorithm that is used in a code signing job.

$sel:encryptionAlgorithm:SigningConfigurationOverrides', signingConfigurationOverrides_encryptionAlgorithm - A specified override of the default encryption algorithm that is used in a code signing job.

signingConfigurationOverrides_hashAlgorithm :: Lens' SigningConfigurationOverrides (Maybe HashAlgorithm)

A specified override of the default hash algorithm that is used in a code signing job.

signingConfigurationOverrides_encryptionAlgorithm :: Lens' SigningConfigurationOverrides (Maybe EncryptionAlgorithm)

A specified override of the default encryption algorithm that is used in a code signing job.

SigningImageFormat

data SigningImageFormat

The image format of a code signing platform or profile.

See: newSigningImageFormat smart constructor.

Constructors

SigningImageFormat'

Fields

supportedFormats :: [ImageFormat]

The supported formats of a code signing image.

defaultFormat :: ImageFormat

The default format of a code signing image.

newSigningImageFormat

Arguments

:: ImageFormat
$sel:defaultFormat:SigningImageFormat'

-> SigningImageFormat

Create a value of SigningImageFormat with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:supportedFormats:SigningImageFormat', signingImageFormat_supportedFormats - The supported formats of a code signing image.

$sel:defaultFormat:SigningImageFormat', signingImageFormat_defaultFormat - The default format of a code signing image.

signingImageFormat_supportedFormats :: Lens' SigningImageFormat [ImageFormat]

The supported formats of a code signing image.

signingImageFormat_defaultFormat :: Lens' SigningImageFormat ImageFormat

The default format of a code signing image.

SigningJob

data SigningJob

Contains information about a signing job.

See: newSigningJob smart constructor.

Constructors

SigningJob'

Fields

status :: Maybe SigningStatus

The status of the signing job.

platformDisplayName :: Maybe Text

The name of a signing platform.

jobId :: Maybe Text

The ID of the signing job.

createdAt :: Maybe POSIX

The date and time that the signing job was created.

signingMaterial :: Maybe SigningMaterial

A SigningMaterial object that contains the Amazon Resource Name (ARN) of the certificate used for the signing job.

isRevoked :: Maybe Bool

Indicates whether the signing job is revoked.

signatureExpiresAt :: Maybe POSIX

The time when the signature of a signing job expires.

profileVersion :: Maybe Text

The version of the signing profile that created a signing job.

profileName :: Maybe Text

The name of the signing profile that created a signing job.

signedObject :: Maybe SignedObject

A SignedObject structure that contains information about a signing job's signed code image.

platformId :: Maybe Text

The unique identifier for a signing platform.

source :: Maybe Source

A Source that contains information about a signing job's code image source.

jobInvoker :: Maybe Text

The AWS account ID of the job invoker.

jobOwner :: Maybe Text

The AWS account ID of the job owner.

newSigningJob :: SigningJob

Create a value of SigningJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SigningJob', signingJob_status - The status of the signing job.

$sel:platformDisplayName:SigningJob', signingJob_platformDisplayName - The name of a signing platform.

$sel:jobId:SigningJob', signingJob_jobId - The ID of the signing job.

$sel:createdAt:SigningJob', signingJob_createdAt - The date and time that the signing job was created.

$sel:signingMaterial:SigningJob', signingJob_signingMaterial - A SigningMaterial object that contains the Amazon Resource Name (ARN) of the certificate used for the signing job.

$sel:isRevoked:SigningJob', signingJob_isRevoked - Indicates whether the signing job is revoked.

$sel:signatureExpiresAt:SigningJob', signingJob_signatureExpiresAt - The time when the signature of a signing job expires.

$sel:profileVersion:SigningJob', signingJob_profileVersion - The version of the signing profile that created a signing job.

$sel:profileName:SigningJob', signingJob_profileName - The name of the signing profile that created a signing job.

$sel:signedObject:SigningJob', signingJob_signedObject - A SignedObject structure that contains information about a signing job's signed code image.

$sel:platformId:SigningJob', signingJob_platformId - The unique identifier for a signing platform.

$sel:source:SigningJob', signingJob_source - A Source that contains information about a signing job's code image source.

$sel:jobInvoker:SigningJob', signingJob_jobInvoker - The AWS account ID of the job invoker.

$sel:jobOwner:SigningJob', signingJob_jobOwner - The AWS account ID of the job owner.

signingJob_status :: Lens' SigningJob (Maybe SigningStatus)

The status of the signing job.

signingJob_platformDisplayName :: Lens' SigningJob (Maybe Text)

The name of a signing platform.

signingJob_jobId :: Lens' SigningJob (Maybe Text)

The ID of the signing job.

signingJob_createdAt :: Lens' SigningJob (Maybe UTCTime)

The date and time that the signing job was created.

signingJob_signingMaterial :: Lens' SigningJob (Maybe SigningMaterial)

A SigningMaterial object that contains the Amazon Resource Name (ARN) of the certificate used for the signing job.

signingJob_isRevoked :: Lens' SigningJob (Maybe Bool)

Indicates whether the signing job is revoked.

signingJob_signatureExpiresAt :: Lens' SigningJob (Maybe UTCTime)

The time when the signature of a signing job expires.

signingJob_profileVersion :: Lens' SigningJob (Maybe Text)

The version of the signing profile that created a signing job.

signingJob_profileName :: Lens' SigningJob (Maybe Text)

The name of the signing profile that created a signing job.

signingJob_signedObject :: Lens' SigningJob (Maybe SignedObject)

A SignedObject structure that contains information about a signing job's signed code image.

signingJob_platformId :: Lens' SigningJob (Maybe Text)

The unique identifier for a signing platform.

signingJob_source :: Lens' SigningJob (Maybe Source)

A Source that contains information about a signing job's code image source.

signingJob_jobInvoker :: Lens' SigningJob (Maybe Text)

The AWS account ID of the job invoker.

signingJob_jobOwner :: Lens' SigningJob (Maybe Text)

The AWS account ID of the job owner.

SigningJobRevocationRecord

data SigningJobRevocationRecord

Revocation information for a signing job.

See: newSigningJobRevocationRecord smart constructor.

Constructors

SigningJobRevocationRecord'

Fields

revokedBy :: Maybe Text

The identity of the revoker.

revokedAt :: Maybe POSIX

The time of revocation.

reason :: Maybe Text

A caller-supplied reason for revocation.

newSigningJobRevocationRecord :: SigningJobRevocationRecord

Create a value of SigningJobRevocationRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revokedBy:SigningJobRevocationRecord', signingJobRevocationRecord_revokedBy - The identity of the revoker.

$sel:revokedAt:SigningJobRevocationRecord', signingJobRevocationRecord_revokedAt - The time of revocation.

$sel:reason:SigningJobRevocationRecord', signingJobRevocationRecord_reason - A caller-supplied reason for revocation.

signingJobRevocationRecord_revokedBy :: Lens' SigningJobRevocationRecord (Maybe Text)

The identity of the revoker.

signingJobRevocationRecord_revokedAt :: Lens' SigningJobRevocationRecord (Maybe UTCTime)

The time of revocation.

signingJobRevocationRecord_reason :: Lens' SigningJobRevocationRecord (Maybe Text)

A caller-supplied reason for revocation.

SigningMaterial

data SigningMaterial

The ACM certificate that is used to sign your code.

See: newSigningMaterial smart constructor.

Constructors

SigningMaterial'

Fields

certificateArn :: Text

The Amazon Resource Name (ARN) of the certificates that is used to sign your code.

newSigningMaterial

Arguments

:: Text
$sel:certificateArn:SigningMaterial'

-> SigningMaterial

Create a value of SigningMaterial with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateArn:SigningMaterial', signingMaterial_certificateArn - The Amazon Resource Name (ARN) of the certificates that is used to sign your code.

signingMaterial_certificateArn :: Lens' SigningMaterial Text

The Amazon Resource Name (ARN) of the certificates that is used to sign your code.

SigningPlatform

data SigningPlatform

Contains information about the signing configurations and parameters that are used to perform a code signing job.

See: newSigningPlatform smart constructor.

Constructors

SigningPlatform'

Fields

category :: Maybe Category

The category of a code signing platform.

signingConfiguration :: Maybe SigningConfiguration

The configuration of a code signing platform. This includes the designated hash algorithm and encryption algorithm of a signing platform.

partner :: Maybe Text

Any partner entities linked to a code signing platform.

revocationSupported :: Maybe Bool

Indicates whether revocation is supported for the platform.

signingImageFormat :: Maybe SigningImageFormat

platformId :: Maybe Text

The ID of a code signing; platform.

displayName :: Maybe Text

The display name of a code signing platform.

maxSizeInMB :: Maybe Int

The maximum size (in MB) of code that can be signed by a code signing platform.

target :: Maybe Text

The types of targets that can be signed by a code signing platform.

newSigningPlatform :: SigningPlatform

Create a value of SigningPlatform with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:category:SigningPlatform', signingPlatform_category - The category of a code signing platform.

$sel:signingConfiguration:SigningPlatform', signingPlatform_signingConfiguration - The configuration of a code signing platform. This includes the designated hash algorithm and encryption algorithm of a signing platform.

$sel:partner:SigningPlatform', signingPlatform_partner - Any partner entities linked to a code signing platform.

$sel:revocationSupported:SigningPlatform', signingPlatform_revocationSupported - Indicates whether revocation is supported for the platform.

$sel:signingImageFormat:SigningPlatform', signingPlatform_signingImageFormat - Undocumented member.

$sel:platformId:SigningPlatform', signingPlatform_platformId - The ID of a code signing; platform.

$sel:displayName:SigningPlatform', signingPlatform_displayName - The display name of a code signing platform.

$sel:maxSizeInMB:SigningPlatform', signingPlatform_maxSizeInMB - The maximum size (in MB) of code that can be signed by a code signing platform.

$sel:target:SigningPlatform', signingPlatform_target - The types of targets that can be signed by a code signing platform.

signingPlatform_category :: Lens' SigningPlatform (Maybe Category)

The category of a code signing platform.

signingPlatform_signingConfiguration :: Lens' SigningPlatform (Maybe SigningConfiguration)

The configuration of a code signing platform. This includes the designated hash algorithm and encryption algorithm of a signing platform.

signingPlatform_partner :: Lens' SigningPlatform (Maybe Text)

Any partner entities linked to a code signing platform.

signingPlatform_revocationSupported :: Lens' SigningPlatform (Maybe Bool)

Indicates whether revocation is supported for the platform.

signingPlatform_signingImageFormat :: Lens' SigningPlatform (Maybe SigningImageFormat)

Undocumented member.

signingPlatform_platformId :: Lens' SigningPlatform (Maybe Text)

The ID of a code signing; platform.

signingPlatform_displayName :: Lens' SigningPlatform (Maybe Text)

The display name of a code signing platform.

signingPlatform_maxSizeInMB :: Lens' SigningPlatform (Maybe Int)

The maximum size (in MB) of code that can be signed by a code signing platform.

signingPlatform_target :: Lens' SigningPlatform (Maybe Text)

The types of targets that can be signed by a code signing platform.

SigningPlatformOverrides

data SigningPlatformOverrides

Any overrides that are applied to the signing configuration of a code signing platform.

See: newSigningPlatformOverrides smart constructor.

Constructors

SigningPlatformOverrides'

Fields

signingConfiguration :: Maybe SigningConfigurationOverrides

A signing configuration that overrides the default encryption or hash algorithm of a signing job.

signingImageFormat :: Maybe ImageFormat

A signed image is a JSON object. When overriding the default signing platform configuration, a customer can select either of two signing formats, JSONEmbedded or JSONDetached. (A third format value, JSON, is reserved for future use.) With JSONEmbedded, the signing image has the payload embedded in it. With JSONDetached, the payload is not be embedded in the signing image.

newSigningPlatformOverrides :: SigningPlatformOverrides

Create a value of SigningPlatformOverrides with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signingConfiguration:SigningPlatformOverrides', signingPlatformOverrides_signingConfiguration - A signing configuration that overrides the default encryption or hash algorithm of a signing job.

$sel:signingImageFormat:SigningPlatformOverrides', signingPlatformOverrides_signingImageFormat - A signed image is a JSON object. When overriding the default signing platform configuration, a customer can select either of two signing formats, JSONEmbedded or JSONDetached. (A third format value, JSON, is reserved for future use.) With JSONEmbedded, the signing image has the payload embedded in it. With JSONDetached, the payload is not be embedded in the signing image.

signingPlatformOverrides_signingConfiguration :: Lens' SigningPlatformOverrides (Maybe SigningConfigurationOverrides)

A signing configuration that overrides the default encryption or hash algorithm of a signing job.

signingPlatformOverrides_signingImageFormat :: Lens' SigningPlatformOverrides (Maybe ImageFormat)

A signed image is a JSON object. When overriding the default signing platform configuration, a customer can select either of two signing formats, JSONEmbedded or JSONDetached. (A third format value, JSON, is reserved for future use.) With JSONEmbedded, the signing image has the payload embedded in it. With JSONDetached, the payload is not be embedded in the signing image.

SigningProfile

data SigningProfile

Contains information about the ACM certificates and code signing configuration parameters that can be used by a given code signing user.

See: newSigningProfile smart constructor.

Constructors

SigningProfile'

Fields

status :: Maybe SigningProfileStatus

The status of a code signing profile.

platformDisplayName :: Maybe Text

The name of the signing platform.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the signing profile.

signingMaterial :: Maybe SigningMaterial

The ACM certificate that is available for use by a signing profile.

profileVersion :: Maybe Text

The version of a signing profile.

profileName :: Maybe Text

The name of the signing profile.

profileVersionArn :: Maybe Text

The ARN of a signing profile, including the profile version.

platformId :: Maybe Text

The ID of a platform that is available for use by a signing profile.

signatureValidityPeriod :: Maybe SignatureValidityPeriod

The validity period for a signing job created using this signing profile.

signingParameters :: Maybe (HashMap Text Text)

The parameters that are available for use by a code signing user.

tags :: Maybe (HashMap Text Text)

A list of tags associated with the signing profile.

newSigningProfile :: SigningProfile

Create a value of SigningProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SigningProfile', signingProfile_status - The status of a code signing profile.

$sel:platformDisplayName:SigningProfile', signingProfile_platformDisplayName - The name of the signing platform.

$sel:arn:SigningProfile', signingProfile_arn - The Amazon Resource Name (ARN) for the signing profile.

$sel:signingMaterial:SigningProfile', signingProfile_signingMaterial - The ACM certificate that is available for use by a signing profile.

$sel:profileVersion:SigningProfile', signingProfile_profileVersion - The version of a signing profile.

$sel:profileName:SigningProfile', signingProfile_profileName - The name of the signing profile.

$sel:profileVersionArn:SigningProfile', signingProfile_profileVersionArn - The ARN of a signing profile, including the profile version.

$sel:platformId:SigningProfile', signingProfile_platformId - The ID of a platform that is available for use by a signing profile.

$sel:signatureValidityPeriod:SigningProfile', signingProfile_signatureValidityPeriod - The validity period for a signing job created using this signing profile.

$sel:signingParameters:SigningProfile', signingProfile_signingParameters - The parameters that are available for use by a code signing user.

$sel:tags:SigningProfile', signingProfile_tags - A list of tags associated with the signing profile.

signingProfile_status :: Lens' SigningProfile (Maybe SigningProfileStatus)

The status of a code signing profile.

signingProfile_platformDisplayName :: Lens' SigningProfile (Maybe Text)

The name of the signing platform.

signingProfile_arn :: Lens' SigningProfile (Maybe Text)

The Amazon Resource Name (ARN) for the signing profile.

signingProfile_signingMaterial :: Lens' SigningProfile (Maybe SigningMaterial)

The ACM certificate that is available for use by a signing profile.

signingProfile_profileVersion :: Lens' SigningProfile (Maybe Text)

The version of a signing profile.

signingProfile_profileName :: Lens' SigningProfile (Maybe Text)

The name of the signing profile.

signingProfile_profileVersionArn :: Lens' SigningProfile (Maybe Text)

The ARN of a signing profile, including the profile version.

signingProfile_platformId :: Lens' SigningProfile (Maybe Text)

The ID of a platform that is available for use by a signing profile.

signingProfile_signatureValidityPeriod :: Lens' SigningProfile (Maybe SignatureValidityPeriod)

The validity period for a signing job created using this signing profile.

signingProfile_signingParameters :: Lens' SigningProfile (Maybe (HashMap Text Text))

The parameters that are available for use by a code signing user.

signingProfile_tags :: Lens' SigningProfile (Maybe (HashMap Text Text))

A list of tags associated with the signing profile.

SigningProfileRevocationRecord

data SigningProfileRevocationRecord

Revocation information for a signing profile.

See: newSigningProfileRevocationRecord smart constructor.

Constructors

SigningProfileRevocationRecord'

Fields

revokedBy :: Maybe Text

The identity of the revoker.

revocationEffectiveFrom :: Maybe POSIX

The time when revocation becomes effective.

revokedAt :: Maybe POSIX

The time when the signing profile was revoked.

newSigningProfileRevocationRecord :: SigningProfileRevocationRecord

Create a value of SigningProfileRevocationRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revokedBy:SigningProfileRevocationRecord', signingProfileRevocationRecord_revokedBy - The identity of the revoker.

$sel:revocationEffectiveFrom:SigningProfileRevocationRecord', signingProfileRevocationRecord_revocationEffectiveFrom - The time when revocation becomes effective.

$sel:revokedAt:SigningProfileRevocationRecord', signingProfileRevocationRecord_revokedAt - The time when the signing profile was revoked.

signingProfileRevocationRecord_revokedBy :: Lens' SigningProfileRevocationRecord (Maybe Text)

The identity of the revoker.

signingProfileRevocationRecord_revocationEffectiveFrom :: Lens' SigningProfileRevocationRecord (Maybe UTCTime)

The time when revocation becomes effective.

signingProfileRevocationRecord_revokedAt :: Lens' SigningProfileRevocationRecord (Maybe UTCTime)

The time when the signing profile was revoked.

Source

data Source

An S3Source object that contains information about the S3 bucket where you saved your unsigned code.

See: newSource smart constructor.

Constructors

Source'

Fields

s3 :: Maybe S3Source

The S3Source object.

newSource :: Source

Create a value of Source with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3:Source', source_s3 - The S3Source object.

source_s3 :: Lens' Source (Maybe S3Source)

The S3Source object.