Amazonka.Signer.Types.S3Destination

Description

Documentation

data S3Destination

The name and prefix of the S3 bucket where code signing saves your signed objects.

See: newS3Destination smart constructor.

Constructors

S3Destination'

Fields

prefix :: Maybe Text

An Amazon S3 prefix that you can use to limit responses to those that begin with the specified prefix.

bucketName :: Maybe Text

Name of the S3 bucket.

newS3Destination :: S3Destination

Create a value of S3Destination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:S3Destination', s3Destination_prefix - An Amazon S3 prefix that you can use to limit responses to those that begin with the specified prefix.

$sel:bucketName:S3Destination', s3Destination_bucketName - Name of the S3 bucket.

s3Destination_prefix :: Lens' S3Destination (Maybe Text)

An Amazon S3 prefix that you can use to limit responses to those that begin with the specified prefix.

s3Destination_bucketName :: Lens' S3Destination (Maybe Text)

Name of the S3 bucket.