Amazonka.Signer.Types.SigningJob

Description

Documentation

data SigningJob

Contains information about a signing job.

See: newSigningJob smart constructor.

Constructors

SigningJob'

Fields

status :: Maybe SigningStatus

The status of the signing job.

platformDisplayName :: Maybe Text

The name of a signing platform.

jobId :: Maybe Text

The ID of the signing job.

createdAt :: Maybe POSIX

The date and time that the signing job was created.

signingMaterial :: Maybe SigningMaterial

A SigningMaterial object that contains the Amazon Resource Name (ARN) of the certificate used for the signing job.

isRevoked :: Maybe Bool

Indicates whether the signing job is revoked.

signatureExpiresAt :: Maybe POSIX

The time when the signature of a signing job expires.

profileVersion :: Maybe Text

The version of the signing profile that created a signing job.

profileName :: Maybe Text

The name of the signing profile that created a signing job.

signedObject :: Maybe SignedObject

A SignedObject structure that contains information about a signing job's signed code image.

platformId :: Maybe Text

The unique identifier for a signing platform.

source :: Maybe Source

A Source that contains information about a signing job's code image source.

jobInvoker :: Maybe Text

The AWS account ID of the job invoker.

jobOwner :: Maybe Text

The AWS account ID of the job owner.

newSigningJob :: SigningJob

Create a value of SigningJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SigningJob', signingJob_status - The status of the signing job.

$sel:platformDisplayName:SigningJob', signingJob_platformDisplayName - The name of a signing platform.

$sel:jobId:SigningJob', signingJob_jobId - The ID of the signing job.

$sel:createdAt:SigningJob', signingJob_createdAt - The date and time that the signing job was created.

$sel:signingMaterial:SigningJob', signingJob_signingMaterial - A SigningMaterial object that contains the Amazon Resource Name (ARN) of the certificate used for the signing job.

$sel:isRevoked:SigningJob', signingJob_isRevoked - Indicates whether the signing job is revoked.

$sel:signatureExpiresAt:SigningJob', signingJob_signatureExpiresAt - The time when the signature of a signing job expires.

$sel:profileVersion:SigningJob', signingJob_profileVersion - The version of the signing profile that created a signing job.

$sel:profileName:SigningJob', signingJob_profileName - The name of the signing profile that created a signing job.

$sel:signedObject:SigningJob', signingJob_signedObject - A SignedObject structure that contains information about a signing job's signed code image.

$sel:platformId:SigningJob', signingJob_platformId - The unique identifier for a signing platform.

$sel:source:SigningJob', signingJob_source - A Source that contains information about a signing job's code image source.

$sel:jobInvoker:SigningJob', signingJob_jobInvoker - The AWS account ID of the job invoker.

$sel:jobOwner:SigningJob', signingJob_jobOwner - The AWS account ID of the job owner.

signingJob_status :: Lens' SigningJob (Maybe SigningStatus)

The status of the signing job.

signingJob_platformDisplayName :: Lens' SigningJob (Maybe Text)

The name of a signing platform.

signingJob_jobId :: Lens' SigningJob (Maybe Text)

The ID of the signing job.

signingJob_createdAt :: Lens' SigningJob (Maybe UTCTime)

The date and time that the signing job was created.

signingJob_signingMaterial :: Lens' SigningJob (Maybe SigningMaterial)

A SigningMaterial object that contains the Amazon Resource Name (ARN) of the certificate used for the signing job.

signingJob_isRevoked :: Lens' SigningJob (Maybe Bool)

Indicates whether the signing job is revoked.

signingJob_signatureExpiresAt :: Lens' SigningJob (Maybe UTCTime)

The time when the signature of a signing job expires.

signingJob_profileVersion :: Lens' SigningJob (Maybe Text)

The version of the signing profile that created a signing job.

signingJob_profileName :: Lens' SigningJob (Maybe Text)

The name of the signing profile that created a signing job.

signingJob_signedObject :: Lens' SigningJob (Maybe SignedObject)

A SignedObject structure that contains information about a signing job's signed code image.

signingJob_platformId :: Lens' SigningJob (Maybe Text)

The unique identifier for a signing platform.

signingJob_source :: Lens' SigningJob (Maybe Source)

A Source that contains information about a signing job's code image source.

signingJob_jobInvoker :: Lens' SigningJob (Maybe Text)

The AWS account ID of the job invoker.

signingJob_jobOwner :: Lens' SigningJob (Maybe Text)

The AWS account ID of the job owner.