Amazonka.Signer.Types.SigningPlatform

Description

Documentation

data SigningPlatform

Contains information about the signing configurations and parameters that are used to perform a code signing job.

See: newSigningPlatform smart constructor.

Constructors

SigningPlatform'

Fields

category :: Maybe Category

The category of a code signing platform.

signingConfiguration :: Maybe SigningConfiguration

The configuration of a code signing platform. This includes the designated hash algorithm and encryption algorithm of a signing platform.

partner :: Maybe Text

Any partner entities linked to a code signing platform.

revocationSupported :: Maybe Bool

Indicates whether revocation is supported for the platform.

signingImageFormat :: Maybe SigningImageFormat

platformId :: Maybe Text

The ID of a code signing; platform.

displayName :: Maybe Text

The display name of a code signing platform.

maxSizeInMB :: Maybe Int

The maximum size (in MB) of code that can be signed by a code signing platform.

target :: Maybe Text

The types of targets that can be signed by a code signing platform.

newSigningPlatform :: SigningPlatform

Create a value of SigningPlatform with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:category:SigningPlatform', signingPlatform_category - The category of a code signing platform.

$sel:signingConfiguration:SigningPlatform', signingPlatform_signingConfiguration - The configuration of a code signing platform. This includes the designated hash algorithm and encryption algorithm of a signing platform.

$sel:partner:SigningPlatform', signingPlatform_partner - Any partner entities linked to a code signing platform.

$sel:revocationSupported:SigningPlatform', signingPlatform_revocationSupported - Indicates whether revocation is supported for the platform.

$sel:signingImageFormat:SigningPlatform', signingPlatform_signingImageFormat - Undocumented member.

$sel:platformId:SigningPlatform', signingPlatform_platformId - The ID of a code signing; platform.

$sel:displayName:SigningPlatform', signingPlatform_displayName - The display name of a code signing platform.

$sel:maxSizeInMB:SigningPlatform', signingPlatform_maxSizeInMB - The maximum size (in MB) of code that can be signed by a code signing platform.

$sel:target:SigningPlatform', signingPlatform_target - The types of targets that can be signed by a code signing platform.

signingPlatform_category :: Lens' SigningPlatform (Maybe Category)

The category of a code signing platform.

signingPlatform_signingConfiguration :: Lens' SigningPlatform (Maybe SigningConfiguration)

The configuration of a code signing platform. This includes the designated hash algorithm and encryption algorithm of a signing platform.

signingPlatform_partner :: Lens' SigningPlatform (Maybe Text)

Any partner entities linked to a code signing platform.

signingPlatform_revocationSupported :: Lens' SigningPlatform (Maybe Bool)

Indicates whether revocation is supported for the platform.

signingPlatform_signingImageFormat :: Lens' SigningPlatform (Maybe SigningImageFormat)

Undocumented member.

signingPlatform_platformId :: Lens' SigningPlatform (Maybe Text)

The ID of a code signing; platform.

signingPlatform_displayName :: Lens' SigningPlatform (Maybe Text)

The display name of a code signing platform.

signingPlatform_maxSizeInMB :: Lens' SigningPlatform (Maybe Int)

The maximum size (in MB) of code that can be signed by a code signing platform.

signingPlatform_target :: Lens' SigningPlatform (Maybe Text)

The types of targets that can be signed by a code signing platform.