Amazonka.Signer.Types.SigningPlatformOverrides

Description

Documentation

data SigningPlatformOverrides

Any overrides that are applied to the signing configuration of a code signing platform.

See: newSigningPlatformOverrides smart constructor.

Constructors

SigningPlatformOverrides'

Fields

signingConfiguration :: Maybe SigningConfigurationOverrides

A signing configuration that overrides the default encryption or hash algorithm of a signing job.

signingImageFormat :: Maybe ImageFormat

A signed image is a JSON object. When overriding the default signing platform configuration, a customer can select either of two signing formats, JSONEmbedded or JSONDetached. (A third format value, JSON, is reserved for future use.) With JSONEmbedded, the signing image has the payload embedded in it. With JSONDetached, the payload is not be embedded in the signing image.

newSigningPlatformOverrides :: SigningPlatformOverrides

Create a value of SigningPlatformOverrides with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signingConfiguration:SigningPlatformOverrides', signingPlatformOverrides_signingConfiguration - A signing configuration that overrides the default encryption or hash algorithm of a signing job.

$sel:signingImageFormat:SigningPlatformOverrides', signingPlatformOverrides_signingImageFormat - A signed image is a JSON object. When overriding the default signing platform configuration, a customer can select either of two signing formats, JSONEmbedded or JSONDetached. (A third format value, JSON, is reserved for future use.) With JSONEmbedded, the signing image has the payload embedded in it. With JSONDetached, the payload is not be embedded in the signing image.

signingPlatformOverrides_signingConfiguration :: Lens' SigningPlatformOverrides (Maybe SigningConfigurationOverrides)

A signing configuration that overrides the default encryption or hash algorithm of a signing job.

signingPlatformOverrides_signingImageFormat :: Lens' SigningPlatformOverrides (Maybe ImageFormat)

A signed image is a JSON object. When overriding the default signing platform configuration, a customer can select either of two signing formats, JSONEmbedded or JSONDetached. (A third format value, JSON, is reserved for future use.) With JSONEmbedded, the signing image has the payload embedded in it. With JSONDetached, the payload is not be embedded in the signing image.