Amazonka.SnowDeviceManagement.DescribeDevice

Description

Checks device-specific information, such as the device type, software version, IP addresses, and lock status.

Creating a Request

data DescribeDevice

See: newDescribeDevice smart constructor.

Constructors

DescribeDevice'

Fields

managedDeviceId :: Text

The ID of the device that you are checking the information of.

newDescribeDevice

Arguments

:: Text
$sel:managedDeviceId:DescribeDevice'

-> DescribeDevice

Create a value of DescribeDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:managedDeviceId:DescribeDevice', describeDevice_managedDeviceId - The ID of the device that you are checking the information of.

Request Lenses

describeDevice_managedDeviceId :: Lens' DescribeDevice Text

The ID of the device that you are checking the information of.

Destructuring the Response

data DescribeDeviceResponse

See: newDescribeDeviceResponse smart constructor.

Constructors

DescribeDeviceResponse'

Fields

deviceState :: Maybe UnlockState

The current state of the device.

deviceCapacities :: Maybe [Capacity]

The hardware specifications of the device.

lastUpdatedAt :: Maybe POSIX

When the device last pushed an update to the Amazon Web Services Cloud. Indicates when the device cache was refreshed.

associatedWithJob :: Maybe Text

The ID of the job used when ordering the device.

lastReachedOutAt :: Maybe POSIX

When the device last contacted the Amazon Web Services Cloud. Indicates that the device is online.

software :: Maybe SoftwareInformation

The software installed on the device.

physicalNetworkInterfaces :: Maybe [PhysicalNetworkInterface]

The network interfaces available on the device.

managedDeviceId :: Maybe Text

The ID of the device that you checked the information for.

managedDeviceArn :: Maybe Text

The Amazon Resource Name (ARN) of the device.

deviceType :: Maybe Text

The type of Amazon Web Services Snow Family device.

tags :: Maybe (HashMap Text Text)

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

httpStatus :: Int

The response's http status code.

newDescribeDeviceResponse

Arguments

:: Int
$sel:httpStatus:DescribeDeviceResponse'

-> DescribeDeviceResponse

Create a value of DescribeDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceState:DescribeDeviceResponse', describeDeviceResponse_deviceState - The current state of the device.

$sel:deviceCapacities:DescribeDeviceResponse', describeDeviceResponse_deviceCapacities - The hardware specifications of the device.

$sel:lastUpdatedAt:DescribeDeviceResponse', describeDeviceResponse_lastUpdatedAt - When the device last pushed an update to the Amazon Web Services Cloud. Indicates when the device cache was refreshed.

$sel:associatedWithJob:DescribeDeviceResponse', describeDeviceResponse_associatedWithJob - The ID of the job used when ordering the device.

$sel:lastReachedOutAt:DescribeDeviceResponse', describeDeviceResponse_lastReachedOutAt - When the device last contacted the Amazon Web Services Cloud. Indicates that the device is online.

$sel:software:DescribeDeviceResponse', describeDeviceResponse_software - The software installed on the device.

$sel:physicalNetworkInterfaces:DescribeDeviceResponse', describeDeviceResponse_physicalNetworkInterfaces - The network interfaces available on the device.

$sel:managedDeviceId:DescribeDevice', describeDeviceResponse_managedDeviceId - The ID of the device that you checked the information for.

$sel:managedDeviceArn:DescribeDeviceResponse', describeDeviceResponse_managedDeviceArn - The Amazon Resource Name (ARN) of the device.

$sel:deviceType:DescribeDeviceResponse', describeDeviceResponse_deviceType - The type of Amazon Web Services Snow Family device.

$sel:tags:DescribeDeviceResponse', describeDeviceResponse_tags - Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

$sel:httpStatus:DescribeDeviceResponse', describeDeviceResponse_httpStatus - The response's http status code.

Response Lenses

describeDeviceResponse_deviceState :: Lens' DescribeDeviceResponse (Maybe UnlockState)

The current state of the device.

describeDeviceResponse_deviceCapacities :: Lens' DescribeDeviceResponse (Maybe [Capacity])

The hardware specifications of the device.

describeDeviceResponse_lastUpdatedAt :: Lens' DescribeDeviceResponse (Maybe UTCTime)

When the device last pushed an update to the Amazon Web Services Cloud. Indicates when the device cache was refreshed.

describeDeviceResponse_associatedWithJob :: Lens' DescribeDeviceResponse (Maybe Text)

The ID of the job used when ordering the device.

describeDeviceResponse_lastReachedOutAt :: Lens' DescribeDeviceResponse (Maybe UTCTime)

When the device last contacted the Amazon Web Services Cloud. Indicates that the device is online.

describeDeviceResponse_software :: Lens' DescribeDeviceResponse (Maybe SoftwareInformation)

The software installed on the device.

describeDeviceResponse_physicalNetworkInterfaces :: Lens' DescribeDeviceResponse (Maybe [PhysicalNetworkInterface])

The network interfaces available on the device.

describeDeviceResponse_managedDeviceId :: Lens' DescribeDeviceResponse (Maybe Text)

The ID of the device that you checked the information for.

describeDeviceResponse_managedDeviceArn :: Lens' DescribeDeviceResponse (Maybe Text)

The Amazon Resource Name (ARN) of the device.

describeDeviceResponse_deviceType :: Lens' DescribeDeviceResponse (Maybe Text)

The type of Amazon Web Services Snow Family device.

describeDeviceResponse_tags :: Lens' DescribeDeviceResponse (Maybe (HashMap Text Text))

Optional metadata that you assign to a resource. You can use tags to categorize a resource in different ways, such as by purpose, owner, or environment.

describeDeviceResponse_httpStatus :: Lens' DescribeDeviceResponse Int

The response's http status code.