Amazonka.Snowball.CreateAddress

Description

Creates an address for a Snow device to be shipped to. In most regions, addresses are validated at the time of creation. The address you provide must be located within the serviceable area of your region. If the address is invalid or unsupported, then an exception is thrown.

Creating a Request

data CreateAddress

See: newCreateAddress smart constructor.

Constructors

CreateAddress'

Fields

address :: Address

The address that you want the Snow device shipped to.

newCreateAddress

Arguments

:: Address
$sel:address:CreateAddress'

-> CreateAddress

Create a value of CreateAddress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:address:CreateAddress', createAddress_address - The address that you want the Snow device shipped to.

Request Lenses

createAddress_address :: Lens' CreateAddress Address

The address that you want the Snow device shipped to.

Destructuring the Response

data CreateAddressResponse

See: newCreateAddressResponse smart constructor.

Constructors

CreateAddressResponse'

Fields

addressId :: Maybe Text

The automatically generated ID for a specific address. You'll use this ID when you create a job to specify which address you want the Snow device for that job shipped to.

httpStatus :: Int

The response's http status code.

newCreateAddressResponse

Arguments

:: Int
$sel:httpStatus:CreateAddressResponse'

-> CreateAddressResponse

Create a value of CreateAddressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addressId:CreateAddressResponse', createAddressResponse_addressId - The automatically generated ID for a specific address. You'll use this ID when you create a job to specify which address you want the Snow device for that job shipped to.

$sel:httpStatus:CreateAddressResponse', createAddressResponse_httpStatus - The response's http status code.

Response Lenses

createAddressResponse_addressId :: Lens' CreateAddressResponse (Maybe Text)

The automatically generated ID for a specific address. You'll use this ID when you create a job to specify which address you want the Snow device for that job shipped to.

createAddressResponse_httpStatus :: Lens' CreateAddressResponse Int

The response's http status code.