Amazonka.Snowball.Lens

Description

Operations

CancelCluster

cancelCluster_clusterId :: Lens' CancelCluster Text

The 39-character ID for the cluster that you want to cancel, for example CID123e4567-e89b-12d3-a456-426655440000.

cancelClusterResponse_httpStatus :: Lens' CancelClusterResponse Int

The response's http status code.

DescribeCluster

describeCluster_clusterId :: Lens' DescribeCluster Text

The automatically generated ID for a cluster.

describeClusterResponse_clusterMetadata :: Lens' DescribeClusterResponse (Maybe ClusterMetadata)

Information about a specific cluster, including shipping information, cluster status, and other important metadata.

describeClusterResponse_httpStatus :: Lens' DescribeClusterResponse Int

The response's http status code.

CreateAddress

createAddress_address :: Lens' CreateAddress Address

The address that you want the Snow device shipped to.

createAddressResponse_addressId :: Lens' CreateAddressResponse (Maybe Text)

The automatically generated ID for a specific address. You'll use this ID when you create a job to specify which address you want the Snow device for that job shipped to.

createAddressResponse_httpStatus :: Lens' CreateAddressResponse Int

The response's http status code.

CreateReturnShippingLabel

createReturnShippingLabel_shippingOption :: Lens' CreateReturnShippingLabel (Maybe ShippingOption)

The shipping speed for a particular job. This speed doesn't dictate how soon the device is returned to AWS. This speed represents how quickly it moves to its destination while in transit. Regional shipping speeds are as follows:

createReturnShippingLabel_jobId :: Lens' CreateReturnShippingLabel Text

The ID for a job that you want to create the return shipping label for; for example, JID123e4567-e89b-12d3-a456-426655440000.

createReturnShippingLabelResponse_status :: Lens' CreateReturnShippingLabelResponse (Maybe ShippingLabelStatus)

The status information of the task on a Snow device that is being returned to AWS.

createReturnShippingLabelResponse_httpStatus :: Lens' CreateReturnShippingLabelResponse Int

The response's http status code.

GetSnowballUsage

getSnowballUsageResponse_snowballsInUse :: Lens' GetSnowballUsageResponse (Maybe Int)

The number of Snow devices that this account is currently using.

getSnowballUsageResponse_snowballLimit :: Lens' GetSnowballUsageResponse (Maybe Int)

The service limit for number of Snow devices this account can have at once. The default service limit is 1 (one).

getSnowballUsageResponse_httpStatus :: Lens' GetSnowballUsageResponse Int

The response's http status code.

DescribeAddresses

describeAddresses_nextToken :: Lens' DescribeAddresses (Maybe Text)

HTTP requests are stateless. To identify what object comes "next" in the list of ADDRESS objects, you have the option of specifying a value for NextToken as the starting point for your list of returned addresses.

describeAddresses_maxResults :: Lens' DescribeAddresses (Maybe Natural)

The number of ADDRESS objects to return.

describeAddressesResponse_addresses :: Lens' DescribeAddressesResponse (Maybe [Address])

The Snow device shipping addresses that were created for this account.

describeAddressesResponse_nextToken :: Lens' DescribeAddressesResponse (Maybe Text)

HTTP requests are stateless. If you use the automatically generated NextToken value in your next DescribeAddresses call, your list of returned addresses will start from this point in the array.

describeAddressesResponse_httpStatus :: Lens' DescribeAddressesResponse Int

The response's http status code.

ListCompatibleImages

listCompatibleImages_nextToken :: Lens' ListCompatibleImages (Maybe Text)

HTTP requests are stateless. To identify what object comes "next" in the list of compatible images, you can specify a value for NextToken as the starting point for your list of returned images.

listCompatibleImages_maxResults :: Lens' ListCompatibleImages (Maybe Natural)

The maximum number of results for the list of compatible images. Currently, a Snowball Edge device can store 10 AMIs.

listCompatibleImagesResponse_compatibleImages :: Lens' ListCompatibleImagesResponse (Maybe [CompatibleImage])

A JSON-formatted object that describes a compatible AMI, including the ID and name for a Snow device AMI.

listCompatibleImagesResponse_nextToken :: Lens' ListCompatibleImagesResponse (Maybe Text)

Because HTTP requests are stateless, this is the starting point for your next list of returned images.

listCompatibleImagesResponse_httpStatus :: Lens' ListCompatibleImagesResponse Int

The response's http status code.

CreateLongTermPricing

createLongTermPricing_snowballType :: Lens' CreateLongTermPricing (Maybe SnowballType)

The type of AWS Snow Family device to use for the long-term pricing job.

createLongTermPricing_isLongTermPricingAutoRenew :: Lens' CreateLongTermPricing (Maybe Bool)

Specifies whether the current long-term pricing type for the device should be renewed.

createLongTermPricing_longTermPricingType :: Lens' CreateLongTermPricing LongTermPricingType

The type of long-term pricing option you want for the device, either 1-year or 3-year long-term pricing.

createLongTermPricingResponse_longTermPricingId :: Lens' CreateLongTermPricingResponse (Maybe Text)

The ID of the long-term pricing type for the device.

createLongTermPricingResponse_httpStatus :: Lens' CreateLongTermPricingResponse Int

The response's http status code.

UpdateCluster

updateCluster_notification :: Lens' UpdateCluster (Maybe Notification)

The new or updated Notification object.

updateCluster_forwardingAddressId :: Lens' UpdateCluster (Maybe Text)

The updated ID for the forwarding address for a cluster. This field is not supported in most regions.

updateCluster_addressId :: Lens' UpdateCluster (Maybe Text)

The ID of the updated Address object.

updateCluster_shippingOption :: Lens' UpdateCluster (Maybe ShippingOption)

The updated shipping option value of this cluster's ShippingDetails object.

updateCluster_resources :: Lens' UpdateCluster (Maybe JobResource)

The updated arrays of JobResource objects that can include updated S3Resource objects or LambdaResource objects.

updateCluster_onDeviceServiceConfiguration :: Lens' UpdateCluster (Maybe OnDeviceServiceConfiguration)

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

updateCluster_description :: Lens' UpdateCluster (Maybe Text)

The updated description of this cluster.

updateCluster_roleARN :: Lens' UpdateCluster (Maybe Text)

The new role Amazon Resource Name (ARN) that you want to associate with this cluster. To create a role ARN, use the CreateRole API action in AWS Identity and Access Management (IAM).

updateCluster_clusterId :: Lens' UpdateCluster Text

The cluster ID of the cluster that you want to update, for example CID123e4567-e89b-12d3-a456-426655440000.

updateClusterResponse_httpStatus :: Lens' UpdateClusterResponse Int

The response's http status code.

GetSoftwareUpdates

getSoftwareUpdates_jobId :: Lens' GetSoftwareUpdates Text

The ID for a job that you want to get the software update file for, for example JID123e4567-e89b-12d3-a456-426655440000.

getSoftwareUpdatesResponse_updatesURI :: Lens' GetSoftwareUpdatesResponse (Maybe Text)

The Amazon S3 presigned URL for the update file associated with the specified JobId value. The software update will be available for 2 days after this request is made. To access an update after the 2 days have passed, you'll have to make another call to GetSoftwareUpdates.

getSoftwareUpdatesResponse_httpStatus :: Lens' GetSoftwareUpdatesResponse Int

The response's http status code.

CreateJob

createJob_jobType :: Lens' CreateJob (Maybe JobType)

Defines the type of job that you're creating.

createJob_kmsKeyARN :: Lens' CreateJob (Maybe Text)

The KmsKeyARN that you want to associate with this job. KmsKeyARNs are created using the CreateKey AWS Key Management Service (KMS) API action.

createJob_remoteManagement :: Lens' CreateJob (Maybe RemoteManagement)

Allows you to securely operate and manage Snowcone devices remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

createJob_notification :: Lens' CreateJob (Maybe Notification)

Defines the Amazon Simple Notification Service (Amazon SNS) notification settings for this job.

createJob_forwardingAddressId :: Lens' CreateJob (Maybe Text)

The forwarding address ID for a job. This field is not supported in most Regions.

createJob_addressId :: Lens' CreateJob (Maybe Text)

The ID for the address that you want the Snow device shipped to.

createJob_snowballType :: Lens' CreateJob (Maybe SnowballType)

The type of AWS Snow Family device to use for this job.

For cluster jobs, AWS Snow Family currently supports only the EDGE device type.

The type of AWS Snow device to use for this job. Currently, the only supported device type for cluster jobs is EDGE.

For more information, see Snowball Edge Device Options in the Snowball Edge Developer Guide.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

createJob_longTermPricingId :: Lens' CreateJob (Maybe Text)

The ID of the long-term pricing type for the device.

createJob_shippingOption :: Lens' CreateJob (Maybe ShippingOption)

The shipping speed for this job. This speed doesn't dictate how soon you'll get the Snow device, rather it represents how quickly the Snow device moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the US, you have access to one-day shipping and two-day shipping.

createJob_resources :: Lens' CreateJob (Maybe JobResource)

Defines the Amazon S3 buckets associated with this job.

With IMPORT jobs, you specify the bucket or buckets that your transferred data will be imported into.

With EXPORT jobs, you specify the bucket or buckets that your transferred data will be exported from. Optionally, you can also specify a KeyRange value. If you choose to export a range, you define the length of the range by providing either an inclusive BeginMarker value, an inclusive EndMarker value, or both. Ranges are UTF-8 binary sorted.

createJob_onDeviceServiceConfiguration :: Lens' CreateJob (Maybe OnDeviceServiceConfiguration)

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

createJob_clusterId :: Lens' CreateJob (Maybe Text)

The ID of a cluster. If you're creating a job for a node in a cluster, you need to provide only this clusterId value. The other job attributes are inherited from the cluster.

createJob_deviceConfiguration :: Lens' CreateJob (Maybe DeviceConfiguration)

Defines the device configuration for an AWS Snowcone job.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

createJob_description :: Lens' CreateJob (Maybe Text)

Defines an optional description of this specific job, for example Important Photos 2016-08-11.

createJob_taxDocuments :: Lens' CreateJob (Maybe TaxDocuments)

The tax documents required in your AWS Region.

createJob_roleARN :: Lens' CreateJob (Maybe Text)

The RoleARN that you want to associate with this job. RoleArns are created using the CreateRole AWS Identity and Access Management (IAM) API action.

createJob_snowballCapacityPreference :: Lens' CreateJob (Maybe SnowballCapacity)

If your job is being created in one of the US regions, you have the option of specifying what size Snow device you'd like for this job. In all other regions, Snowballs come with 80 TB in storage capacity.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

createJobResponse_jobId :: Lens' CreateJobResponse (Maybe Text)

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

createJobResponse_httpStatus :: Lens' CreateJobResponse Int

The response's http status code.

ListLongTermPricing

listLongTermPricing_nextToken :: Lens' ListLongTermPricing (Maybe Text)

Because HTTP requests are stateless, this is the starting point for your next list of ListLongTermPricing to return.

listLongTermPricing_maxResults :: Lens' ListLongTermPricing (Maybe Natural)

The maximum number of ListLongTermPricing objects to return.

listLongTermPricingResponse_longTermPricingEntries :: Lens' ListLongTermPricingResponse (Maybe [LongTermPricingListEntry])

Each LongTermPricingEntry object contains a status, ID, and other information about the LongTermPricing type.

listLongTermPricingResponse_nextToken :: Lens' ListLongTermPricingResponse (Maybe Text)

Because HTTP requests are stateless, this is the starting point for your next list of returned ListLongTermPricing list.

listLongTermPricingResponse_httpStatus :: Lens' ListLongTermPricingResponse Int

The response's http status code.

GetJobManifest

getJobManifest_jobId :: Lens' GetJobManifest Text

The ID for a job that you want to get the manifest file for, for example JID123e4567-e89b-12d3-a456-426655440000.

getJobManifestResponse_manifestURI :: Lens' GetJobManifestResponse (Maybe Text)

The Amazon S3 presigned URL for the manifest file associated with the specified JobId value.

getJobManifestResponse_httpStatus :: Lens' GetJobManifestResponse Int

The response's http status code.

CreateCluster

createCluster_kmsKeyARN :: Lens' CreateCluster (Maybe Text)

The KmsKeyARN value that you want to associate with this cluster. KmsKeyARN values are created by using the CreateKey API action in AWS Key Management Service (AWS KMS).

createCluster_remoteManagement :: Lens' CreateCluster (Maybe RemoteManagement)

Allows you to securely operate and manage Snow devices in a cluster remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

createCluster_notification :: Lens' CreateCluster (Maybe Notification)

The Amazon Simple Notification Service (Amazon SNS) notification settings for this cluster.

createCluster_forwardingAddressId :: Lens' CreateCluster (Maybe Text)

The forwarding address ID for a cluster. This field is not supported in most regions.

createCluster_onDeviceServiceConfiguration :: Lens' CreateCluster (Maybe OnDeviceServiceConfiguration)

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

createCluster_description :: Lens' CreateCluster (Maybe Text)

An optional description of this specific cluster, for example Environmental Data Cluster-01.

createCluster_taxDocuments :: Lens' CreateCluster (Maybe TaxDocuments)

The tax documents required in your AWS Region.

createCluster_jobType :: Lens' CreateCluster JobType

The type of job for this cluster. Currently, the only job type supported for clusters is LOCAL_USE.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

createCluster_resources :: Lens' CreateCluster JobResource

The resources associated with the cluster job. These resources include Amazon S3 buckets and optional AWS Lambda functions written in the Python language.

createCluster_addressId :: Lens' CreateCluster Text

The ID for the address that you want the cluster shipped to.

createCluster_roleARN :: Lens' CreateCluster Text

The RoleARN that you want to associate with this cluster. RoleArn values are created by using the CreateRole API action in AWS Identity and Access Management (IAM).

createCluster_snowballType :: Lens' CreateCluster SnowballType

The type of AWS Snow Family device to use for this cluster.

For cluster jobs, AWS Snow Family currently supports only the EDGE device type.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

createCluster_shippingOption :: Lens' CreateCluster ShippingOption

The shipping speed for each node in this cluster. This speed doesn't dictate how soon you'll get each Snowball Edge device, rather it represents how quickly each device moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the United States of America (US), you have access to one-day shipping and two-day shipping.

In Australia, you have access to express shipping. Typically, devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the US, you have access to one-day shipping and two-day shipping.

createClusterResponse_clusterId :: Lens' CreateClusterResponse (Maybe Text)

The automatically generated ID for a cluster.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.

ListJobs

listJobs_nextToken :: Lens' ListJobs (Maybe Text)

HTTP requests are stateless. To identify what object comes "next" in the list of JobListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

listJobs_maxResults :: Lens' ListJobs (Maybe Natural)

The number of JobListEntry objects to return.

listJobsResponse_jobListEntries :: Lens' ListJobsResponse (Maybe [JobListEntry])

Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs.

listJobsResponse_nextToken :: Lens' ListJobsResponse (Maybe Text)

HTTP requests are stateless. If you use this automatically generated NextToken value in your next ListJobs call, your returned JobListEntry objects will start from this point in the array.

listJobsResponse_httpStatus :: Lens' ListJobsResponse Int

The response's http status code.

UpdateJob

updateJob_notification :: Lens' UpdateJob (Maybe Notification)

The new or updated Notification object.

updateJob_forwardingAddressId :: Lens' UpdateJob (Maybe Text)

The updated ID for the forwarding address for a job. This field is not supported in most regions.

updateJob_addressId :: Lens' UpdateJob (Maybe Text)

The ID of the updated Address object.

updateJob_shippingOption :: Lens' UpdateJob (Maybe ShippingOption)

The updated shipping option value of this job's ShippingDetails object.

updateJob_resources :: Lens' UpdateJob (Maybe JobResource)

The updated JobResource object, or the updated JobResource object.

updateJob_onDeviceServiceConfiguration :: Lens' UpdateJob (Maybe OnDeviceServiceConfiguration)

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

updateJob_description :: Lens' UpdateJob (Maybe Text)

The updated description of this job's JobMetadata object.

updateJob_roleARN :: Lens' UpdateJob (Maybe Text)

The new role Amazon Resource Name (ARN) that you want to associate with this job. To create a role ARN, use the CreateRoleAWS Identity and Access Management (IAM) API action.

updateJob_snowballCapacityPreference :: Lens' UpdateJob (Maybe SnowballCapacity)

The updated SnowballCapacityPreference of this job's JobMetadata object. The 50 TB Snowballs are only available in the US regions.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

updateJob_jobId :: Lens' UpdateJob Text

The job ID of the job that you want to update, for example JID123e4567-e89b-12d3-a456-426655440000.

updateJobResponse_httpStatus :: Lens' UpdateJobResponse Int

The response's http status code.

UpdateJobShipmentState

updateJobShipmentState_jobId :: Lens' UpdateJobShipmentState Text

The job ID of the job whose shipment date you want to update, for example JID123e4567-e89b-12d3-a456-426655440000.

updateJobShipmentState_shipmentState :: Lens' UpdateJobShipmentState ShipmentState

The state of a device when it is being shipped.

Set to RECEIVED when the device arrives at your location.

Set to RETURNED when you have returned the device to AWS.

updateJobShipmentStateResponse_httpStatus :: Lens' UpdateJobShipmentStateResponse Int

The response's http status code.

GetJobUnlockCode

getJobUnlockCode_jobId :: Lens' GetJobUnlockCode Text

The ID for the job that you want to get the UnlockCode value for, for example JID123e4567-e89b-12d3-a456-426655440000.

getJobUnlockCodeResponse_unlockCode :: Lens' GetJobUnlockCodeResponse (Maybe Text)

The UnlockCode value for the specified job. The UnlockCode value can be accessed for up to 360 days after the job has been created.

getJobUnlockCodeResponse_httpStatus :: Lens' GetJobUnlockCodeResponse Int

The response's http status code.

ListClusterJobs

listClusterJobs_nextToken :: Lens' ListClusterJobs (Maybe Text)

HTTP requests are stateless. To identify what object comes "next" in the list of JobListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

listClusterJobs_maxResults :: Lens' ListClusterJobs (Maybe Natural)

The number of JobListEntry objects to return.

listClusterJobs_clusterId :: Lens' ListClusterJobs Text

The 39-character ID for the cluster that you want to list, for example CID123e4567-e89b-12d3-a456-426655440000.

listClusterJobsResponse_jobListEntries :: Lens' ListClusterJobsResponse (Maybe [JobListEntry])

Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs.

listClusterJobsResponse_nextToken :: Lens' ListClusterJobsResponse (Maybe Text)

HTTP requests are stateless. If you use the automatically generated NextToken value in your next ListClusterJobsResult call, your list of returned jobs will start from this point in the array.

listClusterJobsResponse_httpStatus :: Lens' ListClusterJobsResponse Int

The response's http status code.

DescribeJob

describeJob_jobId :: Lens' DescribeJob Text

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

describeJobResponse_jobMetadata :: Lens' DescribeJobResponse (Maybe JobMetadata)

Information about a specific job, including shipping information, job status, and other important metadata.

describeJobResponse_subJobMetadata :: Lens' DescribeJobResponse (Maybe [JobMetadata])

Information about a specific job part (in the case of an export job), including shipping information, job status, and other important metadata.

describeJobResponse_httpStatus :: Lens' DescribeJobResponse Int

The response's http status code.

UpdateLongTermPricing

updateLongTermPricing_isLongTermPricingAutoRenew :: Lens' UpdateLongTermPricing (Maybe Bool)

If set to true, specifies that the current long-term pricing type for the device should be automatically renewed before the long-term pricing contract expires.

updateLongTermPricing_replacementJob :: Lens' UpdateLongTermPricing (Maybe Text)

Specifies that a device that is ordered with long-term pricing should be replaced with a new device.

updateLongTermPricing_longTermPricingId :: Lens' UpdateLongTermPricing Text

The ID of the long-term pricing type for the device.

updateLongTermPricingResponse_httpStatus :: Lens' UpdateLongTermPricingResponse Int

The response's http status code.

ListClusters

listClusters_nextToken :: Lens' ListClusters (Maybe Text)

HTTP requests are stateless. To identify what object comes "next" in the list of ClusterListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

listClusters_maxResults :: Lens' ListClusters (Maybe Natural)

The number of ClusterListEntry objects to return.

listClustersResponse_clusterListEntries :: Lens' ListClustersResponse (Maybe [ClusterListEntry])

Each ClusterListEntry object contains a cluster's state, a cluster's ID, and other important status information.

listClustersResponse_nextToken :: Lens' ListClustersResponse (Maybe Text)

HTTP requests are stateless. If you use the automatically generated NextToken value in your next ClusterListEntry call, your list of returned clusters will start from this point in the array.

listClustersResponse_httpStatus :: Lens' ListClustersResponse Int

The response's http status code.

DescribeAddress

describeAddress_addressId :: Lens' DescribeAddress Text

The automatically generated ID for a specific address.

describeAddressResponse_address :: Lens' DescribeAddressResponse (Maybe Address)

The address that you want the Snow device(s) associated with a specific job to be shipped to.

describeAddressResponse_httpStatus :: Lens' DescribeAddressResponse Int

The response's http status code.

DescribeReturnShippingLabel

describeReturnShippingLabel_jobId :: Lens' DescribeReturnShippingLabel Text

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

describeReturnShippingLabelResponse_status :: Lens' DescribeReturnShippingLabelResponse (Maybe ShippingLabelStatus)

The status information of the task on a Snow device that is being returned to AWS.

describeReturnShippingLabelResponse_expirationDate :: Lens' DescribeReturnShippingLabelResponse (Maybe UTCTime)

The expiration date of the current return shipping label.

describeReturnShippingLabelResponse_httpStatus :: Lens' DescribeReturnShippingLabelResponse Int

The response's http status code.

CancelJob

cancelJob_jobId :: Lens' CancelJob Text

The 39-character job ID for the job that you want to cancel, for example JID123e4567-e89b-12d3-a456-426655440000.

cancelJobResponse_httpStatus :: Lens' CancelJobResponse Int

The response's http status code.

Types

Address

address_isRestricted :: Lens' Address (Maybe Bool)

If the address you are creating is a primary address, then set this option to true. This field is not supported in most regions.

address_street3 :: Lens' Address (Maybe Text)

The third line in a street address that a Snow device is to be delivered to.

address_landmark :: Lens' Address (Maybe Text)

This field is no longer used and the value is ignored.

address_postalCode :: Lens' Address (Maybe Text)

The postal code in an address that a Snow device is to be delivered to.

address_country :: Lens' Address (Maybe Text)

The country in an address that a Snow device is to be delivered to.

address_stateOrProvince :: Lens' Address (Maybe Text)

The state or province in an address that a Snow device is to be delivered to.

address_street2 :: Lens' Address (Maybe Text)

The second line in a street address that a Snow device is to be delivered to.

address_addressId :: Lens' Address (Maybe Text)

The unique ID for an address.

address_city :: Lens' Address (Maybe Text)

The city in an address that a Snow device is to be delivered to.

address_phoneNumber :: Lens' Address (Maybe Text)

The phone number associated with an address that a Snow device is to be delivered to.

address_company :: Lens' Address (Maybe Text)

The name of the company to receive a Snow device at an address.

address_name :: Lens' Address (Maybe Text)

The name of a person to receive a Snow device at an address.

address_prefectureOrDistrict :: Lens' Address (Maybe Text)

This field is no longer used and the value is ignored.

address_street1 :: Lens' Address (Maybe Text)

The first line in a street address that a Snow device is to be delivered to.

ClusterListEntry

clusterListEntry_clusterState :: Lens' ClusterListEntry (Maybe ClusterState)

The current state of this cluster. For information about the state of a specific node, see JobListEntry$JobState.

clusterListEntry_clusterId :: Lens' ClusterListEntry (Maybe Text)

The 39-character ID for the cluster that you want to list, for example CID123e4567-e89b-12d3-a456-426655440000.

clusterListEntry_creationDate :: Lens' ClusterListEntry (Maybe UTCTime)

The creation date for this cluster.

clusterListEntry_description :: Lens' ClusterListEntry (Maybe Text)

Defines an optional description of the cluster, for example Environmental Data Cluster-01.

ClusterMetadata

clusterMetadata_jobType :: Lens' ClusterMetadata (Maybe JobType)

The type of job for this cluster. Currently, the only job type supported for clusters is LOCAL_USE.

clusterMetadata_kmsKeyARN :: Lens' ClusterMetadata (Maybe Text)

The KmsKeyARN Amazon Resource Name (ARN) associated with this cluster. This ARN was created using the CreateKey API action in AWS Key Management Service (AWS KMS).

clusterMetadata_clusterState :: Lens' ClusterMetadata (Maybe ClusterState)

The current status of the cluster.

clusterMetadata_notification :: Lens' ClusterMetadata (Maybe Notification)

The Amazon Simple Notification Service (Amazon SNS) notification settings for this cluster.

clusterMetadata_forwardingAddressId :: Lens' ClusterMetadata (Maybe Text)

The ID of the address that you want a cluster shipped to, after it will be shipped to its primary address. This field is not supported in most regions.

clusterMetadata_addressId :: Lens' ClusterMetadata (Maybe Text)

The automatically generated ID for a specific address.

clusterMetadata_snowballType :: Lens' ClusterMetadata (Maybe SnowballType)

The type of AWS Snow device to use for this cluster.

For cluster jobs, AWS Snow Family currently supports only the EDGE device type.

clusterMetadata_shippingOption :: Lens' ClusterMetadata (Maybe ShippingOption)

The shipping speed for each node in this cluster. This speed doesn't dictate how soon you'll get each device, rather it represents how quickly each device moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the US, you have access to one-day shipping and two-day shipping.

clusterMetadata_resources :: Lens' ClusterMetadata (Maybe JobResource)

The arrays of JobResource objects that can include updated S3Resource objects or LambdaResource objects.

clusterMetadata_onDeviceServiceConfiguration :: Lens' ClusterMetadata (Maybe OnDeviceServiceConfiguration)

Represents metadata and configuration settings for services on an AWS Snow Family device.

clusterMetadata_clusterId :: Lens' ClusterMetadata (Maybe Text)

The automatically generated ID for a cluster.

clusterMetadata_creationDate :: Lens' ClusterMetadata (Maybe UTCTime)

The creation date for this cluster.

clusterMetadata_description :: Lens' ClusterMetadata (Maybe Text)

The optional description of the cluster.

clusterMetadata_taxDocuments :: Lens' ClusterMetadata (Maybe TaxDocuments)

The tax documents required in your AWS Region.

clusterMetadata_roleARN :: Lens' ClusterMetadata (Maybe Text)

The role ARN associated with this cluster. This ARN was created using the CreateRole API action in AWS Identity and Access Management (IAM).

CompatibleImage

compatibleImage_name :: Lens' CompatibleImage (Maybe Text)

The optional name of a compatible image.

compatibleImage_amiId :: Lens' CompatibleImage (Maybe Text)

The unique identifier for an individual Snow device AMI.

DataTransfer

dataTransfer_totalObjects :: Lens' DataTransfer (Maybe Integer)

The total number of objects for a transfer between a Snow device and Amazon S3. This value is set to 0 (zero) until all the keys that will be transferred have been listed.

dataTransfer_totalBytes :: Lens' DataTransfer (Maybe Integer)

The total bytes of data for a transfer between a Snow device and Amazon S3. This value is set to 0 (zero) until all the keys that will be transferred have been listed.

dataTransfer_objectsTransferred :: Lens' DataTransfer (Maybe Integer)

The number of objects transferred between a Snow device and Amazon S3.

dataTransfer_bytesTransferred :: Lens' DataTransfer (Maybe Integer)

The number of bytes transferred between a Snow device and Amazon S3.

DeviceConfiguration

deviceConfiguration_snowconeDeviceConfiguration :: Lens' DeviceConfiguration (Maybe SnowconeDeviceConfiguration)

Returns information about the device configuration for an AWS Snowcone job.

Ec2AmiResource

ec2AmiResource_snowballAmiId :: Lens' Ec2AmiResource (Maybe Text)

The ID of the AMI on the Snow device.

ec2AmiResource_amiId :: Lens' Ec2AmiResource Text

The ID of the AMI in Amazon EC2.

EventTriggerDefinition

eventTriggerDefinition_eventResourceARN :: Lens' EventTriggerDefinition (Maybe Text)

The Amazon Resource Name (ARN) for any local Amazon S3 resource that is an AWS Lambda function's event trigger associated with this job.

INDTaxDocuments

iNDTaxDocuments_gstin :: Lens' INDTaxDocuments (Maybe Text)

The Goods and Services Tax (GST) documents required in AWS Regions in India.

JobListEntry

jobListEntry_jobType :: Lens' JobListEntry (Maybe JobType)

The type of job.

jobListEntry_jobId :: Lens' JobListEntry (Maybe Text)

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

jobListEntry_jobState :: Lens' JobListEntry (Maybe JobState)

The current state of this job.

jobListEntry_snowballType :: Lens' JobListEntry (Maybe SnowballType)

The type of device used with this job.

jobListEntry_creationDate :: Lens' JobListEntry (Maybe UTCTime)

The creation date for this job.

jobListEntry_description :: Lens' JobListEntry (Maybe Text)

The optional description of this specific job, for example Important Photos 2016-08-11.

jobListEntry_isMaster :: Lens' JobListEntry (Maybe Bool)

A value that indicates that this job is a main job. A main job represents a successful request to create an export job. Main jobs aren't associated with any Snowballs. Instead, each main job will have at least one job part, and each job part is associated with a Snowball. It might take some time before the job parts associated with a particular main job are listed, because they are created after the main job is created.

JobLogs

jobLogs_jobFailureLogURI :: Lens' JobLogs (Maybe Text)

A link to an Amazon S3 presigned URL where the job failure log is located.

jobLogs_jobCompletionReportURI :: Lens' JobLogs (Maybe Text)

A link to an Amazon S3 presigned URL where the job completion report is located.

jobLogs_jobSuccessLogURI :: Lens' JobLogs (Maybe Text)

A link to an Amazon S3 presigned URL where the job success log is located.

JobMetadata

jobMetadata_jobType :: Lens' JobMetadata (Maybe JobType)

The type of job.

jobMetadata_kmsKeyARN :: Lens' JobMetadata (Maybe Text)

The Amazon Resource Name (ARN) for the AWS Key Management Service (AWS KMS) key associated with this job. This ARN was created using the CreateKey API action in AWS KMS.

jobMetadata_remoteManagement :: Lens' JobMetadata (Maybe RemoteManagement)

Allows you to securely operate and manage Snowcone devices remotely from outside of your internal network. When set to INSTALLED_AUTOSTART, remote management will automatically be available when the device arrives at your location. Otherwise, you need to use the Snowball Client to manage the device.

jobMetadata_jobId :: Lens' JobMetadata (Maybe Text)

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

jobMetadata_jobLogInfo :: Lens' JobMetadata (Maybe JobLogs)

Links to Amazon S3 presigned URLs for the job report and logs. For import jobs, the PDF job report becomes available at the end of the import process. For export jobs, your job report typically becomes available while the Snow device for your job part is being delivered to you.

jobMetadata_notification :: Lens' JobMetadata (Maybe Notification)

The Amazon Simple Notification Service (Amazon SNS) notification settings associated with a specific job. The Notification object is returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type.

jobMetadata_jobState :: Lens' JobMetadata (Maybe JobState)

The current status of the jobs.

jobMetadata_forwardingAddressId :: Lens' JobMetadata (Maybe Text)

The ID of the address that you want a job shipped to, after it will be shipped to its primary address. This field is not supported in most regions.

jobMetadata_shippingDetails :: Lens' JobMetadata (Maybe ShippingDetails)

A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

jobMetadata_addressId :: Lens' JobMetadata (Maybe Text)

The ID for the address that you want the Snow device shipped to.

jobMetadata_snowballType :: Lens' JobMetadata (Maybe SnowballType)

The type of device used with this job.

jobMetadata_dataTransferProgress :: Lens' JobMetadata (Maybe DataTransfer)

A value that defines the real-time status of a Snow device's data transfer while the device is at AWS. This data is only available while a job has a JobState value of InProgress, for both import and export jobs.

jobMetadata_longTermPricingId :: Lens' JobMetadata (Maybe Text)

The ID of the long-term pricing type for the device.

jobMetadata_resources :: Lens' JobMetadata (Maybe JobResource)

An array of S3Resource objects. Each S3Resource object represents an Amazon S3 bucket that your transferred data will be exported from or imported into.

jobMetadata_onDeviceServiceConfiguration :: Lens' JobMetadata (Maybe OnDeviceServiceConfiguration)

Represents metadata and configuration settings for services on an AWS Snow Family device.

jobMetadata_clusterId :: Lens' JobMetadata (Maybe Text)

The 39-character ID for the cluster, for example CID123e4567-e89b-12d3-a456-426655440000.

jobMetadata_creationDate :: Lens' JobMetadata (Maybe UTCTime)

The creation date for this job.

jobMetadata_deviceConfiguration :: Lens' JobMetadata (Maybe DeviceConfiguration)

Undocumented member.

jobMetadata_description :: Lens' JobMetadata (Maybe Text)

The description of the job, provided at job creation.

jobMetadata_taxDocuments :: Lens' JobMetadata (Maybe TaxDocuments)

The metadata associated with the tax documents required in your AWS Region.

jobMetadata_roleARN :: Lens' JobMetadata (Maybe Text)

The role ARN associated with this job. This ARN was created using the CreateRole API action in AWS Identity and Access Management (IAM).

jobMetadata_snowballCapacityPreference :: Lens' JobMetadata (Maybe SnowballCapacity)

The Snow device capacity preference for this job, specified at job creation. In US regions, you can choose between 50 TB and 80 TB Snowballs. All other regions use 80 TB capacity Snowballs.

For more information, see "https://docs.aws.amazon.com/snowball/latest/snowcone-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide or "https://docs.aws.amazon.com/snowball/latest/developer-guide/snow-device-types.html" (Snow Family Devices and Capacity) in the Snowcone User Guide.

JobResource

jobResource_ec2AmiResources :: Lens' JobResource (Maybe [Ec2AmiResource])

The Amazon Machine Images (AMIs) associated with this job.

jobResource_lambdaResources :: Lens' JobResource (Maybe [LambdaResource])

The Python-language Lambda functions for this job.

jobResource_s3Resources :: Lens' JobResource (Maybe [S3Resource])

An array of S3Resource objects.

KeyRange

keyRange_endMarker :: Lens' KeyRange (Maybe Text)

The key that ends an optional key range for an export job. Ranges are inclusive and UTF-8 binary sorted.

keyRange_beginMarker :: Lens' KeyRange (Maybe Text)

The key that starts an optional key range for an export job. Ranges are inclusive and UTF-8 binary sorted.

LambdaResource

lambdaResource_eventTriggers :: Lens' LambdaResource (Maybe [EventTriggerDefinition])

The array of ARNs for S3Resource objects to trigger the LambdaResource objects associated with this job.

lambdaResource_lambdaArn :: Lens' LambdaResource (Maybe Text)

An Amazon Resource Name (ARN) that represents an AWS Lambda function to be triggered by PUT object actions on the associated local Amazon S3 resource.

LongTermPricingListEntry

longTermPricingListEntry_longTermPricingType :: Lens' LongTermPricingListEntry (Maybe LongTermPricingType)

The type of long-term pricing that was selected for the device.

longTermPricingListEntry_longTermPricingStartDate :: Lens' LongTermPricingListEntry (Maybe UTCTime)

The start date of the long-term pricing contract.

longTermPricingListEntry_snowballType :: Lens' LongTermPricingListEntry (Maybe SnowballType)

The type of AWS Snow Family device associated with this long-term pricing job.

longTermPricingListEntry_longTermPricingId :: Lens' LongTermPricingListEntry (Maybe Text)

The ID of the long-term pricing type for the device.

longTermPricingListEntry_longTermPricingEndDate :: Lens' LongTermPricingListEntry (Maybe UTCTime)

The end date the long-term pricing contract.

longTermPricingListEntry_currentActiveJob :: Lens' LongTermPricingListEntry (Maybe Text)

The current active jobs on the device the long-term pricing type.

longTermPricingListEntry_isLongTermPricingAutoRenew :: Lens' LongTermPricingListEntry (Maybe Bool)

If set to true, specifies that the current long-term pricing type for the device should be automatically renewed before the long-term pricing contract expires.

longTermPricingListEntry_longTermPricingStatus :: Lens' LongTermPricingListEntry (Maybe Text)

The status of the long-term pricing type.

longTermPricingListEntry_jobIds :: Lens' LongTermPricingListEntry (Maybe [Text])

The IDs of the jobs that are associated with a long-term pricing type.

longTermPricingListEntry_replacementJob :: Lens' LongTermPricingListEntry (Maybe Text)

A new device that replaces a device that is ordered with long-term pricing.

NFSOnDeviceServiceConfiguration

nFSOnDeviceServiceConfiguration_storageLimit :: Lens' NFSOnDeviceServiceConfiguration (Maybe Natural)

The maximum NFS storage for one Snowball Family device.

nFSOnDeviceServiceConfiguration_storageUnit :: Lens' NFSOnDeviceServiceConfiguration (Maybe StorageUnit)

The scale unit of the NFS storage on the device.

Valid values: TB.

Notification

notification_notifyAll :: Lens' Notification (Maybe Bool)

Any change in job state will trigger a notification for this job.

notification_snsTopicARN :: Lens' Notification (Maybe Text)

The new SNS TopicArn that you want to associate with this job. You can create Amazon Resource Names (ARNs) for topics by using the CreateTopic Amazon SNS API action.

You can subscribe email addresses to an Amazon SNS topic through the AWS Management Console, or by using the Subscribe Amazon Simple Notification Service (Amazon SNS) API action.

notification_jobStatesToNotify :: Lens' Notification (Maybe [JobState])

The list of job states that will trigger a notification for this job.

OnDeviceServiceConfiguration

onDeviceServiceConfiguration_nFSOnDeviceService :: Lens' OnDeviceServiceConfiguration (Maybe NFSOnDeviceServiceConfiguration)

Represents the NFS service on a Snow Family device.

S3Resource

s3Resource_keyRange :: Lens' S3Resource (Maybe KeyRange)

For export jobs, you can provide an optional KeyRange within a specific Amazon S3 bucket. The length of the range is defined at job creation, and has either an inclusive BeginMarker, an inclusive EndMarker, or both. Ranges are UTF-8 binary sorted.

s3Resource_bucketArn :: Lens' S3Resource (Maybe Text)

The Amazon Resource Name (ARN) of an Amazon S3 bucket.

s3Resource_targetOnDeviceServices :: Lens' S3Resource (Maybe [TargetOnDeviceService])

Specifies the service or services on the Snow Family device that your transferred data will be exported from or imported into. AWS Snow Family supports Amazon S3 and NFS (Network File System).

Shipment

shipment_status :: Lens' Shipment (Maybe Text)

Status information for a shipment.

shipment_trackingNumber :: Lens' Shipment (Maybe Text)

The tracking number for this job. Using this tracking number with your region's carrier's website, you can track a Snow device as the carrier transports it.

For India, the carrier is Amazon Logistics. For all other regions, UPS is the carrier.

ShippingDetails

shippingDetails_shippingOption :: Lens' ShippingDetails (Maybe ShippingOption)

The shipping speed for a particular job. This speed doesn't dictate how soon you'll get the Snow device from the job's creation date. This speed represents how quickly it moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snow devices shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snow devices are delivered in one to seven days.

In the United States of America (US), you have access to one-day shipping and two-day shipping.

shippingDetails_outboundShipment :: Lens' ShippingDetails (Maybe Shipment)

The Status and TrackingNumber values for a Snow device being delivered to the address that you specified for a particular job.

shippingDetails_inboundShipment :: Lens' ShippingDetails (Maybe Shipment)

The Status and TrackingNumber values for a Snow device being returned to AWS for a particular job.

SnowconeDeviceConfiguration

snowconeDeviceConfiguration_wirelessConnection :: Lens' SnowconeDeviceConfiguration (Maybe WirelessConnection)

Configures the wireless connection for the AWS Snowcone device.

TargetOnDeviceService

targetOnDeviceService_transferOption :: Lens' TargetOnDeviceService (Maybe TransferOption)

Specifies whether the data is being imported or exported. You can import or export the data, or use it locally on the device.

targetOnDeviceService_serviceName :: Lens' TargetOnDeviceService (Maybe DeviceServiceName)

Specifies the name of the service on the Snow Family device that your transferred data will be exported from or imported into.

TaxDocuments

taxDocuments_ind :: Lens' TaxDocuments (Maybe INDTaxDocuments)

Undocumented member.

WirelessConnection

wirelessConnection_isWifiEnabled :: Lens' WirelessConnection (Maybe Bool)

Enables the Wi-Fi adapter on an AWS Snowcone device.