Amazonka.Snowball.ListClusterJobs

Description

Returns an array of JobListEntry objects of the specified length. Each JobListEntry object is for a job in the specified cluster and contains a job's state, a job's ID, and other information.

This operation returns paginated results.

Creating a Request

data ListClusterJobs

See: newListClusterJobs smart constructor.

Constructors

ListClusterJobs'

Fields

nextToken :: Maybe Text

HTTP requests are stateless. To identify what object comes "next" in the list of JobListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

maxResults :: Maybe Natural

The number of JobListEntry objects to return.

clusterId :: Text

The 39-character ID for the cluster that you want to list, for example CID123e4567-e89b-12d3-a456-426655440000.

newListClusterJobs

Arguments

:: Text
$sel:clusterId:ListClusterJobs'

-> ListClusterJobs

Create a value of ListClusterJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListClusterJobs', listClusterJobs_nextToken - HTTP requests are stateless. To identify what object comes "next" in the list of JobListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

$sel:maxResults:ListClusterJobs', listClusterJobs_maxResults - The number of JobListEntry objects to return.

$sel:clusterId:ListClusterJobs', listClusterJobs_clusterId - The 39-character ID for the cluster that you want to list, for example CID123e4567-e89b-12d3-a456-426655440000.

Request Lenses

listClusterJobs_nextToken :: Lens' ListClusterJobs (Maybe Text)

HTTP requests are stateless. To identify what object comes "next" in the list of JobListEntry objects, you have the option of specifying NextToken as the starting point for your returned list.

listClusterJobs_maxResults :: Lens' ListClusterJobs (Maybe Natural)

The number of JobListEntry objects to return.

listClusterJobs_clusterId :: Lens' ListClusterJobs Text

The 39-character ID for the cluster that you want to list, for example CID123e4567-e89b-12d3-a456-426655440000.

Destructuring the Response

data ListClusterJobsResponse

See: newListClusterJobsResponse smart constructor.

Constructors

ListClusterJobsResponse'

Fields

jobListEntries :: Maybe [JobListEntry]

Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs.

nextToken :: Maybe Text

HTTP requests are stateless. If you use the automatically generated NextToken value in your next ListClusterJobsResult call, your list of returned jobs will start from this point in the array.

httpStatus :: Int

The response's http status code.

newListClusterJobsResponse

Arguments

:: Int
$sel:httpStatus:ListClusterJobsResponse'

-> ListClusterJobsResponse

Create a value of ListClusterJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobListEntries:ListClusterJobsResponse', listClusterJobsResponse_jobListEntries - Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs.

$sel:nextToken:ListClusterJobs', listClusterJobsResponse_nextToken - HTTP requests are stateless. If you use the automatically generated NextToken value in your next ListClusterJobsResult call, your list of returned jobs will start from this point in the array.

$sel:httpStatus:ListClusterJobsResponse', listClusterJobsResponse_httpStatus - The response's http status code.

Response Lenses

listClusterJobsResponse_jobListEntries :: Lens' ListClusterJobsResponse (Maybe [JobListEntry])

Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of export jobs.

listClusterJobsResponse_nextToken :: Lens' ListClusterJobsResponse (Maybe Text)

HTTP requests are stateless. If you use the automatically generated NextToken value in your next ListClusterJobsResult call, your list of returned jobs will start from this point in the array.

listClusterJobsResponse_httpStatus :: Lens' ListClusterJobsResponse Int

The response's http status code.