Amazonka.Snowball.ListLongTermPricing

Description

Lists all long-term pricing types.

Creating a Request

data ListLongTermPricing

See: newListLongTermPricing smart constructor.

Constructors

ListLongTermPricing'

Fields

nextToken :: Maybe Text

Because HTTP requests are stateless, this is the starting point for your next list of ListLongTermPricing to return.

maxResults :: Maybe Natural

The maximum number of ListLongTermPricing objects to return.

newListLongTermPricing :: ListLongTermPricing

Create a value of ListLongTermPricing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLongTermPricing', listLongTermPricing_nextToken - Because HTTP requests are stateless, this is the starting point for your next list of ListLongTermPricing to return.

$sel:maxResults:ListLongTermPricing', listLongTermPricing_maxResults - The maximum number of ListLongTermPricing objects to return.

Request Lenses

listLongTermPricing_nextToken :: Lens' ListLongTermPricing (Maybe Text)

Because HTTP requests are stateless, this is the starting point for your next list of ListLongTermPricing to return.

listLongTermPricing_maxResults :: Lens' ListLongTermPricing (Maybe Natural)

The maximum number of ListLongTermPricing objects to return.

Destructuring the Response

data ListLongTermPricingResponse

See: newListLongTermPricingResponse smart constructor.

Constructors

ListLongTermPricingResponse'

Fields

longTermPricingEntries :: Maybe [LongTermPricingListEntry]

Each LongTermPricingEntry object contains a status, ID, and other information about the LongTermPricing type.

nextToken :: Maybe Text

Because HTTP requests are stateless, this is the starting point for your next list of returned ListLongTermPricing list.

httpStatus :: Int

The response's http status code.

newListLongTermPricingResponse

Arguments

:: Int
$sel:httpStatus:ListLongTermPricingResponse'

-> ListLongTermPricingResponse

Create a value of ListLongTermPricingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:longTermPricingEntries:ListLongTermPricingResponse', listLongTermPricingResponse_longTermPricingEntries - Each LongTermPricingEntry object contains a status, ID, and other information about the LongTermPricing type.

$sel:nextToken:ListLongTermPricing', listLongTermPricingResponse_nextToken - Because HTTP requests are stateless, this is the starting point for your next list of returned ListLongTermPricing list.

$sel:httpStatus:ListLongTermPricingResponse', listLongTermPricingResponse_httpStatus - The response's http status code.

Response Lenses

listLongTermPricingResponse_longTermPricingEntries :: Lens' ListLongTermPricingResponse (Maybe [LongTermPricingListEntry])

Each LongTermPricingEntry object contains a status, ID, and other information about the LongTermPricing type.

listLongTermPricingResponse_nextToken :: Lens' ListLongTermPricingResponse (Maybe Text)

Because HTTP requests are stateless, this is the starting point for your next list of returned ListLongTermPricing list.

listLongTermPricingResponse_httpStatus :: Lens' ListLongTermPricingResponse Int

The response's http status code.