Amazonka.Snowball.Types.JobListEntry

Description

Documentation

data JobListEntry

Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of an export job.

See: newJobListEntry smart constructor.

Constructors

JobListEntry'

Fields

jobType :: Maybe JobType

The type of job.

jobId :: Maybe Text

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

jobState :: Maybe JobState

The current state of this job.

snowballType :: Maybe SnowballType

The type of device used with this job.

creationDate :: Maybe POSIX

The creation date for this job.

description :: Maybe Text

The optional description of this specific job, for example Important Photos 2016-08-11.

isMaster :: Maybe Bool

A value that indicates that this job is a main job. A main job represents a successful request to create an export job. Main jobs aren't associated with any Snowballs. Instead, each main job will have at least one job part, and each job part is associated with a Snowball. It might take some time before the job parts associated with a particular main job are listed, because they are created after the main job is created.

newJobListEntry :: JobListEntry

Create a value of JobListEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobType:JobListEntry', jobListEntry_jobType - The type of job.

$sel:jobId:JobListEntry', jobListEntry_jobId - The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

$sel:jobState:JobListEntry', jobListEntry_jobState - The current state of this job.

$sel:snowballType:JobListEntry', jobListEntry_snowballType - The type of device used with this job.

$sel:creationDate:JobListEntry', jobListEntry_creationDate - The creation date for this job.

$sel:description:JobListEntry', jobListEntry_description - The optional description of this specific job, for example Important Photos 2016-08-11.

$sel:isMaster:JobListEntry', jobListEntry_isMaster - A value that indicates that this job is a main job. A main job represents a successful request to create an export job. Main jobs aren't associated with any Snowballs. Instead, each main job will have at least one job part, and each job part is associated with a Snowball. It might take some time before the job parts associated with a particular main job are listed, because they are created after the main job is created.

jobListEntry_jobType :: Lens' JobListEntry (Maybe JobType)

The type of job.

jobListEntry_jobId :: Lens' JobListEntry (Maybe Text)

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

jobListEntry_jobState :: Lens' JobListEntry (Maybe JobState)

The current state of this job.

jobListEntry_snowballType :: Lens' JobListEntry (Maybe SnowballType)

The type of device used with this job.

jobListEntry_creationDate :: Lens' JobListEntry (Maybe UTCTime)

The creation date for this job.

jobListEntry_description :: Lens' JobListEntry (Maybe Text)

The optional description of this specific job, for example Important Photos 2016-08-11.

jobListEntry_isMaster :: Lens' JobListEntry (Maybe Bool)

A value that indicates that this job is a main job. A main job represents a successful request to create an export job. Main jobs aren't associated with any Snowballs. Instead, each main job will have at least one job part, and each job part is associated with a Snowball. It might take some time before the job parts associated with a particular main job are listed, because they are created after the main job is created.