Amazonka.Snowball.Types.LongTermPricingListEntry

Description

Documentation

data LongTermPricingListEntry

Each LongTermPricingListEntry object contains information about a long-term pricing type.

See: newLongTermPricingListEntry smart constructor.

Constructors

LongTermPricingListEntry'

Fields

longTermPricingType :: Maybe LongTermPricingType

The type of long-term pricing that was selected for the device.

longTermPricingStartDate :: Maybe POSIX

The start date of the long-term pricing contract.

snowballType :: Maybe SnowballType

The type of AWS Snow Family device associated with this long-term pricing job.

longTermPricingId :: Maybe Text

The ID of the long-term pricing type for the device.

longTermPricingEndDate :: Maybe POSIX

The end date the long-term pricing contract.

currentActiveJob :: Maybe Text

The current active jobs on the device the long-term pricing type.

isLongTermPricingAutoRenew :: Maybe Bool

If set to true, specifies that the current long-term pricing type for the device should be automatically renewed before the long-term pricing contract expires.

longTermPricingStatus :: Maybe Text

The status of the long-term pricing type.

jobIds :: Maybe [Text]

The IDs of the jobs that are associated with a long-term pricing type.

replacementJob :: Maybe Text

A new device that replaces a device that is ordered with long-term pricing.

newLongTermPricingListEntry :: LongTermPricingListEntry

Create a value of LongTermPricingListEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:longTermPricingType:LongTermPricingListEntry', longTermPricingListEntry_longTermPricingType - The type of long-term pricing that was selected for the device.

$sel:longTermPricingStartDate:LongTermPricingListEntry', longTermPricingListEntry_longTermPricingStartDate - The start date of the long-term pricing contract.

$sel:snowballType:LongTermPricingListEntry', longTermPricingListEntry_snowballType - The type of AWS Snow Family device associated with this long-term pricing job.

$sel:longTermPricingId:LongTermPricingListEntry', longTermPricingListEntry_longTermPricingId - The ID of the long-term pricing type for the device.

$sel:longTermPricingEndDate:LongTermPricingListEntry', longTermPricingListEntry_longTermPricingEndDate - The end date the long-term pricing contract.

$sel:currentActiveJob:LongTermPricingListEntry', longTermPricingListEntry_currentActiveJob - The current active jobs on the device the long-term pricing type.

$sel:isLongTermPricingAutoRenew:LongTermPricingListEntry', longTermPricingListEntry_isLongTermPricingAutoRenew - If set to true, specifies that the current long-term pricing type for the device should be automatically renewed before the long-term pricing contract expires.

$sel:longTermPricingStatus:LongTermPricingListEntry', longTermPricingListEntry_longTermPricingStatus - The status of the long-term pricing type.

$sel:jobIds:LongTermPricingListEntry', longTermPricingListEntry_jobIds - The IDs of the jobs that are associated with a long-term pricing type.

$sel:replacementJob:LongTermPricingListEntry', longTermPricingListEntry_replacementJob - A new device that replaces a device that is ordered with long-term pricing.

longTermPricingListEntry_longTermPricingType :: Lens' LongTermPricingListEntry (Maybe LongTermPricingType)

The type of long-term pricing that was selected for the device.

longTermPricingListEntry_longTermPricingStartDate :: Lens' LongTermPricingListEntry (Maybe UTCTime)

The start date of the long-term pricing contract.

longTermPricingListEntry_snowballType :: Lens' LongTermPricingListEntry (Maybe SnowballType)

The type of AWS Snow Family device associated with this long-term pricing job.

longTermPricingListEntry_longTermPricingId :: Lens' LongTermPricingListEntry (Maybe Text)

The ID of the long-term pricing type for the device.

longTermPricingListEntry_longTermPricingEndDate :: Lens' LongTermPricingListEntry (Maybe UTCTime)

The end date the long-term pricing contract.

longTermPricingListEntry_currentActiveJob :: Lens' LongTermPricingListEntry (Maybe Text)

The current active jobs on the device the long-term pricing type.

longTermPricingListEntry_isLongTermPricingAutoRenew :: Lens' LongTermPricingListEntry (Maybe Bool)

If set to true, specifies that the current long-term pricing type for the device should be automatically renewed before the long-term pricing contract expires.

longTermPricingListEntry_longTermPricingStatus :: Lens' LongTermPricingListEntry (Maybe Text)

The status of the long-term pricing type.

longTermPricingListEntry_jobIds :: Lens' LongTermPricingListEntry (Maybe [Text])

The IDs of the jobs that are associated with a long-term pricing type.

longTermPricingListEntry_replacementJob :: Lens' LongTermPricingListEntry (Maybe Text)

A new device that replaces a device that is ordered with long-term pricing.