Amazonka.Snowball.UpdateLongTermPricing

Description

Updates the long-term pricing type.

Creating a Request

data UpdateLongTermPricing

See: newUpdateLongTermPricing smart constructor.

Constructors

UpdateLongTermPricing'

Fields

isLongTermPricingAutoRenew :: Maybe Bool

If set to true, specifies that the current long-term pricing type for the device should be automatically renewed before the long-term pricing contract expires.

replacementJob :: Maybe Text

Specifies that a device that is ordered with long-term pricing should be replaced with a new device.

longTermPricingId :: Text

The ID of the long-term pricing type for the device.

newUpdateLongTermPricing

Arguments

:: Text
$sel:longTermPricingId:UpdateLongTermPricing'

-> UpdateLongTermPricing

Create a value of UpdateLongTermPricing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isLongTermPricingAutoRenew:UpdateLongTermPricing', updateLongTermPricing_isLongTermPricingAutoRenew - If set to true, specifies that the current long-term pricing type for the device should be automatically renewed before the long-term pricing contract expires.

$sel:replacementJob:UpdateLongTermPricing', updateLongTermPricing_replacementJob - Specifies that a device that is ordered with long-term pricing should be replaced with a new device.

$sel:longTermPricingId:UpdateLongTermPricing', updateLongTermPricing_longTermPricingId - The ID of the long-term pricing type for the device.

Request Lenses

updateLongTermPricing_isLongTermPricingAutoRenew :: Lens' UpdateLongTermPricing (Maybe Bool)

If set to true, specifies that the current long-term pricing type for the device should be automatically renewed before the long-term pricing contract expires.

updateLongTermPricing_replacementJob :: Lens' UpdateLongTermPricing (Maybe Text)

Specifies that a device that is ordered with long-term pricing should be replaced with a new device.

updateLongTermPricing_longTermPricingId :: Lens' UpdateLongTermPricing Text

The ID of the long-term pricing type for the device.

Destructuring the Response

data UpdateLongTermPricingResponse

See: newUpdateLongTermPricingResponse smart constructor.

Constructors

UpdateLongTermPricingResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateLongTermPricingResponse

Arguments

:: Int
$sel:httpStatus:UpdateLongTermPricingResponse'

-> UpdateLongTermPricingResponse

Create a value of UpdateLongTermPricingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateLongTermPricingResponse', updateLongTermPricingResponse_httpStatus - The response's http status code.

Response Lenses

updateLongTermPricingResponse_httpStatus :: Lens' UpdateLongTermPricingResponse Int

The response's http status code.