Amazonka.StepFunctions.DescribeActivity

Description

Describes an activity.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

Creating a Request

data DescribeActivity

See: newDescribeActivity smart constructor.

Constructors

DescribeActivity'

Fields

activityArn :: Text

The Amazon Resource Name (ARN) of the activity to describe.

newDescribeActivity

Arguments

:: Text
$sel:activityArn:DescribeActivity'

-> DescribeActivity

Create a value of DescribeActivity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activityArn:DescribeActivity', describeActivity_activityArn - The Amazon Resource Name (ARN) of the activity to describe.

Request Lenses

describeActivity_activityArn :: Lens' DescribeActivity Text

The Amazon Resource Name (ARN) of the activity to describe.

Destructuring the Response

data DescribeActivityResponse

See: newDescribeActivityResponse smart constructor.

Constructors

DescribeActivityResponse'

Fields

httpStatus :: Int

The response's http status code.

activityArn :: Text

The Amazon Resource Name (ARN) that identifies the activity.

name :: Text

The name of the activity.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

creationDate :: POSIX

The date the activity is created.

newDescribeActivityResponse

Arguments

:: Int
$sel:httpStatus:DescribeActivityResponse'

-> Text
$sel:activityArn:DescribeActivity'

-> Text
$sel:name:DescribeActivityResponse'

-> UTCTime
$sel:creationDate:DescribeActivityResponse'

-> DescribeActivityResponse

Create a value of DescribeActivityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeActivityResponse', describeActivityResponse_httpStatus - The response's http status code.

$sel:activityArn:DescribeActivity', describeActivityResponse_activityArn - The Amazon Resource Name (ARN) that identifies the activity.

$sel:name:DescribeActivityResponse', describeActivityResponse_name - The name of the activity.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

$sel:creationDate:DescribeActivityResponse', describeActivityResponse_creationDate - The date the activity is created.

Response Lenses

describeActivityResponse_httpStatus :: Lens' DescribeActivityResponse Int

The response's http status code.

describeActivityResponse_activityArn :: Lens' DescribeActivityResponse Text

The Amazon Resource Name (ARN) that identifies the activity.

describeActivityResponse_name :: Lens' DescribeActivityResponse Text

The name of the activity.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

describeActivityResponse_creationDate :: Lens' DescribeActivityResponse UTCTime

The date the activity is created.