Amazonka.StepFunctions.DescribeStateMachineForExecution

Description

Describes the state machine associated with a specific execution.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

This API action is not supported by EXPRESS state machines.

Creating a Request

data DescribeStateMachineForExecution

See: newDescribeStateMachineForExecution smart constructor.

Constructors

DescribeStateMachineForExecution'

Fields

executionArn :: Text

The Amazon Resource Name (ARN) of the execution you want state machine information for.

newDescribeStateMachineForExecution

Arguments

:: Text
$sel:executionArn:DescribeStateMachineForExecution'

-> DescribeStateMachineForExecution

Create a value of DescribeStateMachineForExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionArn:DescribeStateMachineForExecution', describeStateMachineForExecution_executionArn - The Amazon Resource Name (ARN) of the execution you want state machine information for.

Request Lenses

describeStateMachineForExecution_executionArn :: Lens' DescribeStateMachineForExecution Text

The Amazon Resource Name (ARN) of the execution you want state machine information for.

Destructuring the Response

data DescribeStateMachineForExecutionResponse

See: newDescribeStateMachineForExecutionResponse smart constructor.

Constructors

DescribeStateMachineForExecutionResponse'

Fields

tracingConfiguration :: Maybe TracingConfiguration

Selects whether AWS X-Ray tracing is enabled.

loggingConfiguration :: Maybe LoggingConfiguration

httpStatus :: Int

The response's http status code.

stateMachineArn :: Text

The Amazon Resource Name (ARN) of the state machine associated with the execution.

name :: Text

The name of the state machine associated with the execution.

definition :: Sensitive Text

The Amazon States Language definition of the state machine. See Amazon States Language.

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM role of the State Machine for the execution.

updateDate :: POSIX

The date and time the state machine associated with an execution was updated. For a newly created state machine, this is the creation date.

newDescribeStateMachineForExecutionResponse

Arguments

:: Int
$sel:httpStatus:DescribeStateMachineForExecutionResponse'

-> Text
$sel:stateMachineArn:DescribeStateMachineForExecutionResponse'

-> Text
$sel:name:DescribeStateMachineForExecutionResponse'

-> Text
$sel:definition:DescribeStateMachineForExecutionResponse'

-> Text
$sel:roleArn:DescribeStateMachineForExecutionResponse'

-> UTCTime
$sel:updateDate:DescribeStateMachineForExecutionResponse'

-> DescribeStateMachineForExecutionResponse

Create a value of DescribeStateMachineForExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tracingConfiguration:DescribeStateMachineForExecutionResponse', describeStateMachineForExecutionResponse_tracingConfiguration - Selects whether AWS X-Ray tracing is enabled.

$sel:loggingConfiguration:DescribeStateMachineForExecutionResponse', describeStateMachineForExecutionResponse_loggingConfiguration - Undocumented member.

$sel:httpStatus:DescribeStateMachineForExecutionResponse', describeStateMachineForExecutionResponse_httpStatus - The response's http status code.

$sel:stateMachineArn:DescribeStateMachineForExecutionResponse', describeStateMachineForExecutionResponse_stateMachineArn - The Amazon Resource Name (ARN) of the state machine associated with the execution.

$sel:name:DescribeStateMachineForExecutionResponse', describeStateMachineForExecutionResponse_name - The name of the state machine associated with the execution.

$sel:definition:DescribeStateMachineForExecutionResponse', describeStateMachineForExecutionResponse_definition - The Amazon States Language definition of the state machine. See Amazon States Language.

$sel:roleArn:DescribeStateMachineForExecutionResponse', describeStateMachineForExecutionResponse_roleArn - The Amazon Resource Name (ARN) of the IAM role of the State Machine for the execution.

$sel:updateDate:DescribeStateMachineForExecutionResponse', describeStateMachineForExecutionResponse_updateDate - The date and time the state machine associated with an execution was updated. For a newly created state machine, this is the creation date.

Response Lenses

describeStateMachineForExecutionResponse_tracingConfiguration :: Lens' DescribeStateMachineForExecutionResponse (Maybe TracingConfiguration)

Selects whether AWS X-Ray tracing is enabled.

describeStateMachineForExecutionResponse_loggingConfiguration :: Lens' DescribeStateMachineForExecutionResponse (Maybe LoggingConfiguration)

Undocumented member.

describeStateMachineForExecutionResponse_httpStatus :: Lens' DescribeStateMachineForExecutionResponse Int

The response's http status code.

describeStateMachineForExecutionResponse_stateMachineArn :: Lens' DescribeStateMachineForExecutionResponse Text

The Amazon Resource Name (ARN) of the state machine associated with the execution.

describeStateMachineForExecutionResponse_name :: Lens' DescribeStateMachineForExecutionResponse Text

The name of the state machine associated with the execution.

describeStateMachineForExecutionResponse_definition :: Lens' DescribeStateMachineForExecutionResponse Text

The Amazon States Language definition of the state machine. See Amazon States Language.

describeStateMachineForExecutionResponse_roleArn :: Lens' DescribeStateMachineForExecutionResponse Text

The Amazon Resource Name (ARN) of the IAM role of the State Machine for the execution.

describeStateMachineForExecutionResponse_updateDate :: Lens' DescribeStateMachineForExecutionResponse UTCTime

The date and time the state machine associated with an execution was updated. For a newly created state machine, this is the creation date.