Amazonka.StepFunctions.GetExecutionHistory

Description

Returns the history of the specified execution as a list of events. By default, the results are returned in ascending order of the timeStamp of the events. Use the reverseOrder parameter to get the latest events first.

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

This API action is not supported by EXPRESS state machines.

This operation returns paginated results.

Creating a Request

data GetExecutionHistory

See: newGetExecutionHistory smart constructor.

Constructors

GetExecutionHistory'

Fields

reverseOrder :: Maybe Bool

Lists events in descending order of their timeStamp.

includeExecutionData :: Maybe Bool

You can select whether execution data (input or output of a history event) is returned. The default is true.

nextToken :: Maybe Text

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

maxResults :: Maybe Natural

The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default.

This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

executionArn :: Text

The Amazon Resource Name (ARN) of the execution.

newGetExecutionHistory

Arguments

:: Text
$sel:executionArn:GetExecutionHistory'

-> GetExecutionHistory

Create a value of GetExecutionHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reverseOrder:GetExecutionHistory', getExecutionHistory_reverseOrder - Lists events in descending order of their timeStamp.

$sel:includeExecutionData:GetExecutionHistory', getExecutionHistory_includeExecutionData - You can select whether execution data (input or output of a history event) is returned. The default is true.

$sel:nextToken:GetExecutionHistory', getExecutionHistory_nextToken - If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

$sel:maxResults:GetExecutionHistory', getExecutionHistory_maxResults - The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default.

This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

$sel:executionArn:GetExecutionHistory', getExecutionHistory_executionArn - The Amazon Resource Name (ARN) of the execution.

Request Lenses

getExecutionHistory_reverseOrder :: Lens' GetExecutionHistory (Maybe Bool)

Lists events in descending order of their timeStamp.

getExecutionHistory_includeExecutionData :: Lens' GetExecutionHistory (Maybe Bool)

You can select whether execution data (input or output of a history event) is returned. The default is true.

getExecutionHistory_nextToken :: Lens' GetExecutionHistory (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

getExecutionHistory_maxResults :: Lens' GetExecutionHistory (Maybe Natural)

The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default.

This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

getExecutionHistory_executionArn :: Lens' GetExecutionHistory Text

The Amazon Resource Name (ARN) of the execution.

Destructuring the Response

data GetExecutionHistoryResponse

See: newGetExecutionHistoryResponse smart constructor.

Constructors

GetExecutionHistoryResponse'

Fields

nextToken :: Maybe Text

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

httpStatus :: Int

The response's http status code.

events :: [HistoryEvent]

The list of events that occurred in the execution.

newGetExecutionHistoryResponse

Arguments

:: Int
$sel:httpStatus:GetExecutionHistoryResponse'

-> GetExecutionHistoryResponse

Create a value of GetExecutionHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetExecutionHistory', getExecutionHistoryResponse_nextToken - If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

$sel:httpStatus:GetExecutionHistoryResponse', getExecutionHistoryResponse_httpStatus - The response's http status code.

$sel:events:GetExecutionHistoryResponse', getExecutionHistoryResponse_events - The list of events that occurred in the execution.

Response Lenses

getExecutionHistoryResponse_nextToken :: Lens' GetExecutionHistoryResponse (Maybe Text)

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

getExecutionHistoryResponse_httpStatus :: Lens' GetExecutionHistoryResponse Int

The response's http status code.

getExecutionHistoryResponse_events :: Lens' GetExecutionHistoryResponse [HistoryEvent]

The list of events that occurred in the execution.