Amazonka.StepFunctions.StartExecution

Description

Starts a state machine execution.

StartExecution is idempotent. If StartExecution is called with the same name and input as a running execution, the call will succeed and return the same response as the original request. If the execution is closed or if the input is different, it will return a 400 ExecutionAlreadyExists error. Names can be reused after 90 days.

Creating a Request

data StartExecution

See: newStartExecution smart constructor.

Constructors

StartExecution'

Fields

input :: Maybe (Sensitive Text)

The string that contains the JSON input data for the execution, for example:

"input": "{\"first_name\" : \"test\"}"

If you don't include any JSON input data, you still must include the two braces, for example: "input": "{}"

Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

name :: Maybe Text

The name of the execution. This name must be unique for your AWS account, region, and state machine for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

traceHeader :: Maybe Text

Passes the AWS X-Ray trace header. The trace header can also be passed in the request payload.

stateMachineArn :: Text

The Amazon Resource Name (ARN) of the state machine to execute.

newStartExecution

Arguments

:: Text
$sel:stateMachineArn:StartExecution'

-> StartExecution

Create a value of StartExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:input:StartExecution', startExecution_input - The string that contains the JSON input data for the execution, for example:

"input": "{\"first_name\" : \"test\"}"

If you don't include any JSON input data, you still must include the two braces, for example: "input": "{}"

Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

$sel:name:StartExecution', startExecution_name - The name of the execution. This name must be unique for your AWS account, region, and state machine for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

$sel:traceHeader:StartExecution', startExecution_traceHeader - Passes the AWS X-Ray trace header. The trace header can also be passed in the request payload.

$sel:stateMachineArn:StartExecution', startExecution_stateMachineArn - The Amazon Resource Name (ARN) of the state machine to execute.

Request Lenses

startExecution_input :: Lens' StartExecution (Maybe Text)

The string that contains the JSON input data for the execution, for example:

"input": "{\"first_name\" : \"test\"}"

If you don't include any JSON input data, you still must include the two braces, for example: "input": "{}"

Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

startExecution_name :: Lens' StartExecution (Maybe Text)

The name of the execution. This name must be unique for your AWS account, region, and state machine for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide.

A name must not contain:

white space

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ ` $ & , ; : /

control characters (U+0000-001F, U+007F-009F)

To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

startExecution_traceHeader :: Lens' StartExecution (Maybe Text)

Passes the AWS X-Ray trace header. The trace header can also be passed in the request payload.

startExecution_stateMachineArn :: Lens' StartExecution Text

The Amazon Resource Name (ARN) of the state machine to execute.

Destructuring the Response

data StartExecutionResponse

See: newStartExecutionResponse smart constructor.

Constructors

StartExecutionResponse'

Fields

httpStatus :: Int

The response's http status code.

executionArn :: Text

The Amazon Resource Name (ARN) that identifies the execution.

startDate :: POSIX

The date the execution is started.

newStartExecutionResponse

Arguments

:: Int
$sel:httpStatus:StartExecutionResponse'

-> Text
$sel:executionArn:StartExecutionResponse'

-> UTCTime
$sel:startDate:StartExecutionResponse'

-> StartExecutionResponse

Create a value of StartExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartExecutionResponse', startExecutionResponse_httpStatus - The response's http status code.

$sel:executionArn:StartExecutionResponse', startExecutionResponse_executionArn - The Amazon Resource Name (ARN) that identifies the execution.

$sel:startDate:StartExecutionResponse', startExecutionResponse_startDate - The date the execution is started.

Response Lenses

startExecutionResponse_httpStatus :: Lens' StartExecutionResponse Int

The response's http status code.

startExecutionResponse_executionArn :: Lens' StartExecutionResponse Text

The Amazon Resource Name (ARN) that identifies the execution.

startExecutionResponse_startDate :: Lens' StartExecutionResponse UTCTime

The date the execution is started.