Amazonka.StepFunctions.StartSyncExecution

Description

Starts a Synchronous Express state machine execution.

Creating a Request

data StartSyncExecution

See: newStartSyncExecution smart constructor.

Constructors

StartSyncExecution'

Fields

input :: Maybe (Sensitive Text)

The string that contains the JSON input data for the execution, for example:

"input": "{\"first_name\" : \"test\"}"

If you don't include any JSON input data, you still must include the two braces, for example: "input": "{}"

Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

name :: Maybe Text

The name of the execution.

traceHeader :: Maybe Text

Passes the AWS X-Ray trace header. The trace header can also be passed in the request payload.

stateMachineArn :: Text

The Amazon Resource Name (ARN) of the state machine to execute.

newStartSyncExecution

Arguments

:: Text
$sel:stateMachineArn:StartSyncExecution'

-> StartSyncExecution

Create a value of StartSyncExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:input:StartSyncExecution', startSyncExecution_input - The string that contains the JSON input data for the execution, for example:

"input": "{\"first_name\" : \"test\"}"

If you don't include any JSON input data, you still must include the two braces, for example: "input": "{}"

Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

$sel:name:StartSyncExecution', startSyncExecution_name - The name of the execution.

$sel:traceHeader:StartSyncExecution', startSyncExecution_traceHeader - Passes the AWS X-Ray trace header. The trace header can also be passed in the request payload.

$sel:stateMachineArn:StartSyncExecution', startSyncExecution_stateMachineArn - The Amazon Resource Name (ARN) of the state machine to execute.

Request Lenses

startSyncExecution_input :: Lens' StartSyncExecution (Maybe Text)

The string that contains the JSON input data for the execution, for example:

"input": "{\"first_name\" : \"test\"}"

If you don't include any JSON input data, you still must include the two braces, for example: "input": "{}"

Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

startSyncExecution_name :: Lens' StartSyncExecution (Maybe Text)

The name of the execution.

startSyncExecution_traceHeader :: Lens' StartSyncExecution (Maybe Text)

Passes the AWS X-Ray trace header. The trace header can also be passed in the request payload.

startSyncExecution_stateMachineArn :: Lens' StartSyncExecution Text

The Amazon Resource Name (ARN) of the state machine to execute.

Destructuring the Response

data StartSyncExecutionResponse

See: newStartSyncExecutionResponse smart constructor.

Constructors

StartSyncExecutionResponse'

Fields

inputDetails :: Maybe CloudWatchEventsExecutionDataDetails

error :: Maybe (Sensitive Text)

The error code of the failure.

input :: Maybe (Sensitive Text)

The string that contains the JSON input data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

cause :: Maybe (Sensitive Text)

A more detailed explanation of the cause of the failure.

name :: Maybe Text

The name of the execution.

stateMachineArn :: Maybe Text

The Amazon Resource Name (ARN) that identifies the state machine.

output :: Maybe (Sensitive Text)

The JSON output data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

This field is set only if the execution succeeds. If the execution fails, this field is null.

outputDetails :: Maybe CloudWatchEventsExecutionDataDetails

traceHeader :: Maybe Text

The AWS X-Ray trace header that was passed to the execution.

billingDetails :: Maybe BillingDetails

An object that describes workflow billing details, including billed duration and memory use.

httpStatus :: Int

The response's http status code.

executionArn :: Text

The Amazon Resource Name (ARN) that identifies the execution.

startDate :: POSIX

The date the execution is started.

stopDate :: POSIX

If the execution has already ended, the date the execution stopped.

status :: SyncExecutionStatus

The current status of the execution.

newStartSyncExecutionResponse

Arguments

:: Int
$sel:httpStatus:StartSyncExecutionResponse'

-> Text
$sel:executionArn:StartSyncExecutionResponse'

-> UTCTime
$sel:startDate:StartSyncExecutionResponse'

-> UTCTime
$sel:stopDate:StartSyncExecutionResponse'

-> SyncExecutionStatus
$sel:status:StartSyncExecutionResponse'

-> StartSyncExecutionResponse

Create a value of StartSyncExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputDetails:StartSyncExecutionResponse', startSyncExecutionResponse_inputDetails - Undocumented member.

$sel:error:StartSyncExecutionResponse', startSyncExecutionResponse_error - The error code of the failure.

$sel:input:StartSyncExecution', startSyncExecutionResponse_input - The string that contains the JSON input data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

$sel:cause:StartSyncExecutionResponse', startSyncExecutionResponse_cause - A more detailed explanation of the cause of the failure.

$sel:name:StartSyncExecution', startSyncExecutionResponse_name - The name of the execution.

$sel:stateMachineArn:StartSyncExecution', startSyncExecutionResponse_stateMachineArn - The Amazon Resource Name (ARN) that identifies the state machine.

$sel:output:StartSyncExecutionResponse', startSyncExecutionResponse_output - The JSON output data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

This field is set only if the execution succeeds. If the execution fails, this field is null.

$sel:outputDetails:StartSyncExecutionResponse', startSyncExecutionResponse_outputDetails - Undocumented member.

$sel:traceHeader:StartSyncExecution', startSyncExecutionResponse_traceHeader - The AWS X-Ray trace header that was passed to the execution.

$sel:billingDetails:StartSyncExecutionResponse', startSyncExecutionResponse_billingDetails - An object that describes workflow billing details, including billed duration and memory use.

$sel:httpStatus:StartSyncExecutionResponse', startSyncExecutionResponse_httpStatus - The response's http status code.

$sel:executionArn:StartSyncExecutionResponse', startSyncExecutionResponse_executionArn - The Amazon Resource Name (ARN) that identifies the execution.

$sel:startDate:StartSyncExecutionResponse', startSyncExecutionResponse_startDate - The date the execution is started.

$sel:stopDate:StartSyncExecutionResponse', startSyncExecutionResponse_stopDate - If the execution has already ended, the date the execution stopped.

$sel:status:StartSyncExecutionResponse', startSyncExecutionResponse_status - The current status of the execution.

Response Lenses

startSyncExecutionResponse_inputDetails :: Lens' StartSyncExecutionResponse (Maybe CloudWatchEventsExecutionDataDetails)

Undocumented member.

startSyncExecutionResponse_error :: Lens' StartSyncExecutionResponse (Maybe Text)

The error code of the failure.

startSyncExecutionResponse_input :: Lens' StartSyncExecutionResponse (Maybe Text)

The string that contains the JSON input data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

startSyncExecutionResponse_cause :: Lens' StartSyncExecutionResponse (Maybe Text)

A more detailed explanation of the cause of the failure.

startSyncExecutionResponse_name :: Lens' StartSyncExecutionResponse (Maybe Text)

The name of the execution.

startSyncExecutionResponse_stateMachineArn :: Lens' StartSyncExecutionResponse (Maybe Text)

The Amazon Resource Name (ARN) that identifies the state machine.

startSyncExecutionResponse_output :: Lens' StartSyncExecutionResponse (Maybe Text)

The JSON output data of the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

This field is set only if the execution succeeds. If the execution fails, this field is null.

startSyncExecutionResponse_outputDetails :: Lens' StartSyncExecutionResponse (Maybe CloudWatchEventsExecutionDataDetails)

Undocumented member.

startSyncExecutionResponse_traceHeader :: Lens' StartSyncExecutionResponse (Maybe Text)

The AWS X-Ray trace header that was passed to the execution.

startSyncExecutionResponse_billingDetails :: Lens' StartSyncExecutionResponse (Maybe BillingDetails)

An object that describes workflow billing details, including billed duration and memory use.

startSyncExecutionResponse_httpStatus :: Lens' StartSyncExecutionResponse Int

The response's http status code.

startSyncExecutionResponse_executionArn :: Lens' StartSyncExecutionResponse Text

The Amazon Resource Name (ARN) that identifies the execution.

startSyncExecutionResponse_startDate :: Lens' StartSyncExecutionResponse UTCTime

The date the execution is started.

startSyncExecutionResponse_stopDate :: Lens' StartSyncExecutionResponse UTCTime

If the execution has already ended, the date the execution stopped.

startSyncExecutionResponse_status :: Lens' StartSyncExecutionResponse SyncExecutionStatus

The current status of the execution.