Amazonka.StepFunctions.Types.ActivityScheduledEventDetails

Description

Documentation

data ActivityScheduledEventDetails

Contains details about an activity scheduled during an execution.

See: newActivityScheduledEventDetails smart constructor.

Constructors

ActivityScheduledEventDetails'

Fields

heartbeatInSeconds :: Maybe Integer

The maximum allowed duration between two heartbeats for the activity task.

inputDetails :: Maybe HistoryEventExecutionDataDetails

Contains details about the input for an execution history event.

input :: Maybe (Sensitive Text)

The JSON data input to the activity task. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

timeoutInSeconds :: Maybe Integer

The maximum allowed duration of the activity task.

resource :: Text

The Amazon Resource Name (ARN) of the scheduled activity.

newActivityScheduledEventDetails

Arguments

:: Text
$sel:resource:ActivityScheduledEventDetails'

-> ActivityScheduledEventDetails

Create a value of ActivityScheduledEventDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:heartbeatInSeconds:ActivityScheduledEventDetails', activityScheduledEventDetails_heartbeatInSeconds - The maximum allowed duration between two heartbeats for the activity task.

$sel:inputDetails:ActivityScheduledEventDetails', activityScheduledEventDetails_inputDetails - Contains details about the input for an execution history event.

$sel:input:ActivityScheduledEventDetails', activityScheduledEventDetails_input - The JSON data input to the activity task. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

$sel:timeoutInSeconds:ActivityScheduledEventDetails', activityScheduledEventDetails_timeoutInSeconds - The maximum allowed duration of the activity task.

$sel:resource:ActivityScheduledEventDetails', activityScheduledEventDetails_resource - The Amazon Resource Name (ARN) of the scheduled activity.

activityScheduledEventDetails_heartbeatInSeconds :: Lens' ActivityScheduledEventDetails (Maybe Integer)

The maximum allowed duration between two heartbeats for the activity task.

activityScheduledEventDetails_inputDetails :: Lens' ActivityScheduledEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about the input for an execution history event.

activityScheduledEventDetails_input :: Lens' ActivityScheduledEventDetails (Maybe Text)

The JSON data input to the activity task. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

activityScheduledEventDetails_timeoutInSeconds :: Lens' ActivityScheduledEventDetails (Maybe Integer)

The maximum allowed duration of the activity task.

activityScheduledEventDetails_resource :: Lens' ActivityScheduledEventDetails Text

The Amazon Resource Name (ARN) of the scheduled activity.