Amazonka.StepFunctions.Types.ExecutionStartedEventDetails

Description

Documentation

data ExecutionStartedEventDetails

Contains details about the start of the execution.

See: newExecutionStartedEventDetails smart constructor.

Constructors

ExecutionStartedEventDetails'

Fields

inputDetails :: Maybe HistoryEventExecutionDataDetails

Contains details about the input for an execution history event.

input :: Maybe (Sensitive Text)

The JSON data input to the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role used for executing AWS Lambda tasks.

newExecutionStartedEventDetails :: ExecutionStartedEventDetails

Create a value of ExecutionStartedEventDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputDetails:ExecutionStartedEventDetails', executionStartedEventDetails_inputDetails - Contains details about the input for an execution history event.

$sel:input:ExecutionStartedEventDetails', executionStartedEventDetails_input - The JSON data input to the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

$sel:roleArn:ExecutionStartedEventDetails', executionStartedEventDetails_roleArn - The Amazon Resource Name (ARN) of the IAM role used for executing AWS Lambda tasks.

executionStartedEventDetails_inputDetails :: Lens' ExecutionStartedEventDetails (Maybe HistoryEventExecutionDataDetails)

Contains details about the input for an execution history event.

executionStartedEventDetails_input :: Lens' ExecutionStartedEventDetails (Maybe Text)

The JSON data input to the execution. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

executionStartedEventDetails_roleArn :: Lens' ExecutionStartedEventDetails (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used for executing AWS Lambda tasks.