Amazonka.StepFunctions.Types.ExecutionTimedOutEventDetails

Description

Documentation

data ExecutionTimedOutEventDetails

Contains details about the execution timeout that occurred during the execution.

See: newExecutionTimedOutEventDetails smart constructor.

Constructors

ExecutionTimedOutEventDetails'

Fields

error :: Maybe (Sensitive Text)

The error code of the failure.

cause :: Maybe (Sensitive Text)

A more detailed explanation of the cause of the timeout.

newExecutionTimedOutEventDetails :: ExecutionTimedOutEventDetails

Create a value of ExecutionTimedOutEventDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:error:ExecutionTimedOutEventDetails', executionTimedOutEventDetails_error - The error code of the failure.

$sel:cause:ExecutionTimedOutEventDetails', executionTimedOutEventDetails_cause - A more detailed explanation of the cause of the timeout.

executionTimedOutEventDetails_error :: Lens' ExecutionTimedOutEventDetails (Maybe Text)

The error code of the failure.

executionTimedOutEventDetails_cause :: Lens' ExecutionTimedOutEventDetails (Maybe Text)

A more detailed explanation of the cause of the timeout.