Amazonka.StepFunctions.Types.HistoryEvent

Description

Documentation

data HistoryEvent

Contains details about the events of an execution.

See: newHistoryEvent smart constructor.

Constructors

HistoryEvent'

Fields

mapStateStartedEventDetails :: Maybe MapStateStartedEventDetails

Contains details about Map state that was started.

taskSubmitFailedEventDetails :: Maybe TaskSubmitFailedEventDetails

Contains details about a task that where the submit failed.

taskStartedEventDetails :: Maybe TaskStartedEventDetails

Contains details about a task that was started.

activityStartedEventDetails :: Maybe ActivityStartedEventDetails

taskSubmittedEventDetails :: Maybe TaskSubmittedEventDetails

Contains details about a submitted task.

lambdaFunctionStartFailedEventDetails :: Maybe LambdaFunctionStartFailedEventDetails

Contains details about a lambda function that failed to start during an execution.

taskStartFailedEventDetails :: Maybe TaskStartFailedEventDetails

Contains details about a task that failed to start.

stateExitedEventDetails :: Maybe StateExitedEventDetails

lambdaFunctionSucceededEventDetails :: Maybe LambdaFunctionSucceededEventDetails

Contains details about a lambda function that terminated successfully during an execution.

taskSucceededEventDetails :: Maybe TaskSucceededEventDetails

Contains details about a task that succeeded.

activitySucceededEventDetails :: Maybe ActivitySucceededEventDetails

mapIterationAbortedEventDetails :: Maybe MapIterationEventDetails

Contains details about an iteration of a Map state that was aborted.

mapIterationSucceededEventDetails :: Maybe MapIterationEventDetails

Contains details about an iteration of a Map state that succeeded.

mapIterationStartedEventDetails :: Maybe MapIterationEventDetails

Contains details about an iteration of a Map state that was started.

lambdaFunctionTimedOutEventDetails :: Maybe LambdaFunctionTimedOutEventDetails

taskTimedOutEventDetails :: Maybe TaskTimedOutEventDetails

Contains details about a task that timed out.

activityTimedOutEventDetails :: Maybe ActivityTimedOutEventDetails

executionFailedEventDetails :: Maybe ExecutionFailedEventDetails

executionAbortedEventDetails :: Maybe ExecutionAbortedEventDetails

executionSucceededEventDetails :: Maybe ExecutionSucceededEventDetails

lambdaFunctionScheduledEventDetails :: Maybe LambdaFunctionScheduledEventDetails

taskScheduledEventDetails :: Maybe TaskScheduledEventDetails

Contains details about a task that was scheduled.

activityScheduledEventDetails :: Maybe ActivityScheduledEventDetails

executionStartedEventDetails :: Maybe ExecutionStartedEventDetails

activityScheduleFailedEventDetails :: Maybe ActivityScheduleFailedEventDetails

Contains details about an activity schedule event that failed during an execution.

lambdaFunctionScheduleFailedEventDetails :: Maybe LambdaFunctionScheduleFailedEventDetails

stateEnteredEventDetails :: Maybe StateEnteredEventDetails

previousEventId :: Maybe Integer

The id of the previous event.

activityFailedEventDetails :: Maybe ActivityFailedEventDetails

taskFailedEventDetails :: Maybe TaskFailedEventDetails

Contains details about the failure of a task.

lambdaFunctionFailedEventDetails :: Maybe LambdaFunctionFailedEventDetails

executionTimedOutEventDetails :: Maybe ExecutionTimedOutEventDetails

mapIterationFailedEventDetails :: Maybe MapIterationEventDetails

Contains details about an iteration of a Map state that failed.

timestamp :: POSIX

The date and time the event occurred.

type' :: HistoryEventType

The type of the event.

id :: Integer

The id of the event. Events are numbered sequentially, starting at one.

newHistoryEvent

Arguments

:: UTCTime
$sel:timestamp:HistoryEvent'

-> HistoryEventType
$sel:type':HistoryEvent'

-> Integer
$sel:id:HistoryEvent'

-> HistoryEvent

Create a value of HistoryEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mapStateStartedEventDetails:HistoryEvent', historyEvent_mapStateStartedEventDetails - Contains details about Map state that was started.

$sel:taskSubmitFailedEventDetails:HistoryEvent', historyEvent_taskSubmitFailedEventDetails - Contains details about a task that where the submit failed.

$sel:taskStartedEventDetails:HistoryEvent', historyEvent_taskStartedEventDetails - Contains details about a task that was started.

$sel:activityStartedEventDetails:HistoryEvent', historyEvent_activityStartedEventDetails - Undocumented member.

$sel:taskSubmittedEventDetails:HistoryEvent', historyEvent_taskSubmittedEventDetails - Contains details about a submitted task.

$sel:lambdaFunctionStartFailedEventDetails:HistoryEvent', historyEvent_lambdaFunctionStartFailedEventDetails - Contains details about a lambda function that failed to start during an execution.

$sel:taskStartFailedEventDetails:HistoryEvent', historyEvent_taskStartFailedEventDetails - Contains details about a task that failed to start.

$sel:stateExitedEventDetails:HistoryEvent', historyEvent_stateExitedEventDetails - Undocumented member.

$sel:lambdaFunctionSucceededEventDetails:HistoryEvent', historyEvent_lambdaFunctionSucceededEventDetails - Contains details about a lambda function that terminated successfully during an execution.

$sel:taskSucceededEventDetails:HistoryEvent', historyEvent_taskSucceededEventDetails - Contains details about a task that succeeded.

$sel:activitySucceededEventDetails:HistoryEvent', historyEvent_activitySucceededEventDetails - Undocumented member.

$sel:mapIterationAbortedEventDetails:HistoryEvent', historyEvent_mapIterationAbortedEventDetails - Contains details about an iteration of a Map state that was aborted.

$sel:mapIterationSucceededEventDetails:HistoryEvent', historyEvent_mapIterationSucceededEventDetails - Contains details about an iteration of a Map state that succeeded.

$sel:mapIterationStartedEventDetails:HistoryEvent', historyEvent_mapIterationStartedEventDetails - Contains details about an iteration of a Map state that was started.

$sel:lambdaFunctionTimedOutEventDetails:HistoryEvent', historyEvent_lambdaFunctionTimedOutEventDetails - Undocumented member.

$sel:taskTimedOutEventDetails:HistoryEvent', historyEvent_taskTimedOutEventDetails - Contains details about a task that timed out.

$sel:activityTimedOutEventDetails:HistoryEvent', historyEvent_activityTimedOutEventDetails - Undocumented member.

$sel:executionFailedEventDetails:HistoryEvent', historyEvent_executionFailedEventDetails - Undocumented member.

$sel:executionAbortedEventDetails:HistoryEvent', historyEvent_executionAbortedEventDetails - Undocumented member.

$sel:executionSucceededEventDetails:HistoryEvent', historyEvent_executionSucceededEventDetails - Undocumented member.

$sel:lambdaFunctionScheduledEventDetails:HistoryEvent', historyEvent_lambdaFunctionScheduledEventDetails - Undocumented member.

$sel:taskScheduledEventDetails:HistoryEvent', historyEvent_taskScheduledEventDetails - Contains details about a task that was scheduled.

$sel:activityScheduledEventDetails:HistoryEvent', historyEvent_activityScheduledEventDetails - Undocumented member.

$sel:executionStartedEventDetails:HistoryEvent', historyEvent_executionStartedEventDetails - Undocumented member.

$sel:activityScheduleFailedEventDetails:HistoryEvent', historyEvent_activityScheduleFailedEventDetails - Contains details about an activity schedule event that failed during an execution.

$sel:lambdaFunctionScheduleFailedEventDetails:HistoryEvent', historyEvent_lambdaFunctionScheduleFailedEventDetails - Undocumented member.

$sel:stateEnteredEventDetails:HistoryEvent', historyEvent_stateEnteredEventDetails - Undocumented member.

$sel:previousEventId:HistoryEvent', historyEvent_previousEventId - The id of the previous event.

$sel:activityFailedEventDetails:HistoryEvent', historyEvent_activityFailedEventDetails - Undocumented member.

$sel:taskFailedEventDetails:HistoryEvent', historyEvent_taskFailedEventDetails - Contains details about the failure of a task.

$sel:lambdaFunctionFailedEventDetails:HistoryEvent', historyEvent_lambdaFunctionFailedEventDetails - Undocumented member.

$sel:executionTimedOutEventDetails:HistoryEvent', historyEvent_executionTimedOutEventDetails - Undocumented member.

$sel:mapIterationFailedEventDetails:HistoryEvent', historyEvent_mapIterationFailedEventDetails - Contains details about an iteration of a Map state that failed.

$sel:timestamp:HistoryEvent', historyEvent_timestamp - The date and time the event occurred.

$sel:type':HistoryEvent', historyEvent_type - The type of the event.

$sel:id:HistoryEvent', historyEvent_id - The id of the event. Events are numbered sequentially, starting at one.

historyEvent_mapStateStartedEventDetails :: Lens' HistoryEvent (Maybe MapStateStartedEventDetails)

Contains details about Map state that was started.

historyEvent_taskSubmitFailedEventDetails :: Lens' HistoryEvent (Maybe TaskSubmitFailedEventDetails)

Contains details about a task that where the submit failed.

historyEvent_taskStartedEventDetails :: Lens' HistoryEvent (Maybe TaskStartedEventDetails)

Contains details about a task that was started.

historyEvent_activityStartedEventDetails :: Lens' HistoryEvent (Maybe ActivityStartedEventDetails)

Undocumented member.

historyEvent_taskSubmittedEventDetails :: Lens' HistoryEvent (Maybe TaskSubmittedEventDetails)

Contains details about a submitted task.

historyEvent_lambdaFunctionStartFailedEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionStartFailedEventDetails)

Contains details about a lambda function that failed to start during an execution.

historyEvent_taskStartFailedEventDetails :: Lens' HistoryEvent (Maybe TaskStartFailedEventDetails)

Contains details about a task that failed to start.

historyEvent_stateExitedEventDetails :: Lens' HistoryEvent (Maybe StateExitedEventDetails)

Undocumented member.

historyEvent_lambdaFunctionSucceededEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionSucceededEventDetails)

Contains details about a lambda function that terminated successfully during an execution.

historyEvent_taskSucceededEventDetails :: Lens' HistoryEvent (Maybe TaskSucceededEventDetails)

Contains details about a task that succeeded.

historyEvent_activitySucceededEventDetails :: Lens' HistoryEvent (Maybe ActivitySucceededEventDetails)

Undocumented member.

historyEvent_mapIterationAbortedEventDetails :: Lens' HistoryEvent (Maybe MapIterationEventDetails)

Contains details about an iteration of a Map state that was aborted.

historyEvent_mapIterationSucceededEventDetails :: Lens' HistoryEvent (Maybe MapIterationEventDetails)

Contains details about an iteration of a Map state that succeeded.

historyEvent_mapIterationStartedEventDetails :: Lens' HistoryEvent (Maybe MapIterationEventDetails)

Contains details about an iteration of a Map state that was started.

historyEvent_lambdaFunctionTimedOutEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionTimedOutEventDetails)

Undocumented member.

historyEvent_taskTimedOutEventDetails :: Lens' HistoryEvent (Maybe TaskTimedOutEventDetails)

Contains details about a task that timed out.

historyEvent_activityTimedOutEventDetails :: Lens' HistoryEvent (Maybe ActivityTimedOutEventDetails)

Undocumented member.

historyEvent_executionFailedEventDetails :: Lens' HistoryEvent (Maybe ExecutionFailedEventDetails)

Undocumented member.

historyEvent_executionAbortedEventDetails :: Lens' HistoryEvent (Maybe ExecutionAbortedEventDetails)

Undocumented member.

historyEvent_executionSucceededEventDetails :: Lens' HistoryEvent (Maybe ExecutionSucceededEventDetails)

Undocumented member.

historyEvent_lambdaFunctionScheduledEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionScheduledEventDetails)

Undocumented member.

historyEvent_taskScheduledEventDetails :: Lens' HistoryEvent (Maybe TaskScheduledEventDetails)

Contains details about a task that was scheduled.

historyEvent_activityScheduledEventDetails :: Lens' HistoryEvent (Maybe ActivityScheduledEventDetails)

Undocumented member.

historyEvent_executionStartedEventDetails :: Lens' HistoryEvent (Maybe ExecutionStartedEventDetails)

Undocumented member.

historyEvent_activityScheduleFailedEventDetails :: Lens' HistoryEvent (Maybe ActivityScheduleFailedEventDetails)

Contains details about an activity schedule event that failed during an execution.

historyEvent_lambdaFunctionScheduleFailedEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionScheduleFailedEventDetails)

Undocumented member.

historyEvent_stateEnteredEventDetails :: Lens' HistoryEvent (Maybe StateEnteredEventDetails)

Undocumented member.

historyEvent_previousEventId :: Lens' HistoryEvent (Maybe Integer)

The id of the previous event.

historyEvent_activityFailedEventDetails :: Lens' HistoryEvent (Maybe ActivityFailedEventDetails)

Undocumented member.

historyEvent_taskFailedEventDetails :: Lens' HistoryEvent (Maybe TaskFailedEventDetails)

Contains details about the failure of a task.

historyEvent_lambdaFunctionFailedEventDetails :: Lens' HistoryEvent (Maybe LambdaFunctionFailedEventDetails)

Undocumented member.

historyEvent_executionTimedOutEventDetails :: Lens' HistoryEvent (Maybe ExecutionTimedOutEventDetails)

Undocumented member.

historyEvent_mapIterationFailedEventDetails :: Lens' HistoryEvent (Maybe MapIterationEventDetails)

Contains details about an iteration of a Map state that failed.

historyEvent_timestamp :: Lens' HistoryEvent UTCTime

The date and time the event occurred.

historyEvent_type :: Lens' HistoryEvent HistoryEventType

The type of the event.

historyEvent_id :: Lens' HistoryEvent Integer

The id of the event. Events are numbered sequentially, starting at one.