Amazonka.StepFunctions.Types.TaskScheduledEventDetails

Description

Documentation

data TaskScheduledEventDetails

Contains details about a task scheduled during an execution.

See: newTaskScheduledEventDetails smart constructor.

Constructors

TaskScheduledEventDetails'

Fields

heartbeatInSeconds :: Maybe Integer

The maximum allowed duration between two heartbeats for the task.

timeoutInSeconds :: Maybe Integer

The maximum allowed duration of the task.

resourceType :: Text

The action of the resource called by a task state.

resource :: Text

The service name of the resource in a task state.

region :: Text

The region of the scheduled task

parameters :: Sensitive Text

The JSON data passed to the resource referenced in a task state. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

newTaskScheduledEventDetails

Arguments

:: Text
$sel:resourceType:TaskScheduledEventDetails'

-> Text
$sel:resource:TaskScheduledEventDetails'

-> Text
$sel:region:TaskScheduledEventDetails'

-> Text
$sel:parameters:TaskScheduledEventDetails'

-> TaskScheduledEventDetails

Create a value of TaskScheduledEventDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:heartbeatInSeconds:TaskScheduledEventDetails', taskScheduledEventDetails_heartbeatInSeconds - The maximum allowed duration between two heartbeats for the task.

$sel:timeoutInSeconds:TaskScheduledEventDetails', taskScheduledEventDetails_timeoutInSeconds - The maximum allowed duration of the task.

$sel:resourceType:TaskScheduledEventDetails', taskScheduledEventDetails_resourceType - The action of the resource called by a task state.

$sel:resource:TaskScheduledEventDetails', taskScheduledEventDetails_resource - The service name of the resource in a task state.

$sel:region:TaskScheduledEventDetails', taskScheduledEventDetails_region - The region of the scheduled task

$sel:parameters:TaskScheduledEventDetails', taskScheduledEventDetails_parameters - The JSON data passed to the resource referenced in a task state. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

taskScheduledEventDetails_heartbeatInSeconds :: Lens' TaskScheduledEventDetails (Maybe Integer)

The maximum allowed duration between two heartbeats for the task.

taskScheduledEventDetails_timeoutInSeconds :: Lens' TaskScheduledEventDetails (Maybe Integer)

The maximum allowed duration of the task.

taskScheduledEventDetails_resourceType :: Lens' TaskScheduledEventDetails Text

The action of the resource called by a task state.

taskScheduledEventDetails_resource :: Lens' TaskScheduledEventDetails Text

The service name of the resource in a task state.

taskScheduledEventDetails_region :: Lens' TaskScheduledEventDetails Text

The region of the scheduled task

taskScheduledEventDetails_parameters :: Lens' TaskScheduledEventDetails Text

The JSON data passed to the resource referenced in a task state. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.