Amazonka.StorageGateway.ActivateGateway

Description

Activates the gateway you previously deployed on your host. In the activation process, you specify information such as the Amazon Web Services Region that you want to use for storing snapshots or tapes, the time zone for scheduled snapshots the gateway snapshot schedule window, an activation key, and a name for your gateway. The activation process also associates your gateway with your account. For more information, see UpdateGatewayInformation.

You must turn on the gateway VM before you can activate your gateway.

Creating a Request

data ActivateGateway

A JSON object containing one or more of the following fields:

ActivateGatewayInput$ActivationKey

ActivateGatewayInput$GatewayName

ActivateGatewayInput$GatewayRegion

ActivateGatewayInput$GatewayTimezone

ActivateGatewayInput$GatewayType

ActivateGatewayInput$MediumChangerType

ActivateGatewayInput$TapeDriveType

See: newActivateGateway smart constructor.

Constructors

ActivateGateway'

Fields

mediumChangerType :: Maybe Text

The value that indicates the type of medium changer to use for tape gateway. This field is optional.

Valid Values: STK-L700 | AWS-Gateway-VTL | IBM-03584L32-0402

tapeDriveType :: Maybe Text

The value that indicates the type of tape drive to use for tape gateway. This field is optional.

Valid Values: IBM-ULT3580-TD5

gatewayType :: Maybe Text

A value that defines the type of gateway to activate. The type specified is critical to all later functions of the gateway and cannot be changed after activation. The default value is CACHED.

Valid Values: STORED | CACHED | VTL | FILE_S3 | FILE_FSX_SMB|

tags :: Maybe [Tag]

A list of up to 50 tags that you can assign to the gateway. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers that can be represented in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256 characters.

activationKey :: Text

Your gateway activation key. You can obtain the activation key by sending an HTTP GET request with redirects enabled to the gateway IP address (port 80). The redirect URL returned in the response provides you the activation key for your gateway in the query string parameter activationKey. It may also include other activation-related parameters, however, these are merely defaults -- the arguments you pass to the ActivateGateway API call determine the actual configuration of your gateway.

For more information, see Getting activation key in the Storage Gateway User Guide.

gatewayName :: Text

The name you configured for your gateway.

gatewayTimezone :: Text

A value that indicates the time zone you want to set for the gateway. The time zone is of the format "GMT-hr:mm" or "GMT+hr:mm". For example, GMT-4:00 indicates the time is 4 hours behind GMT. GMT+2:00 indicates the time is 2 hours ahead of GMT. The time zone is used, for example, for scheduling snapshots and your gateway's maintenance schedule.

gatewayRegion :: Text

A value that indicates the Amazon Web Services Region where you want to store your data. The gateway Amazon Web Services Region specified must be the same Amazon Web Services Region as the Amazon Web Services Region in your Host header in the request. For more information about available Amazon Web Services Regions and endpoints for Storage Gateway, see Storage Gateway endpoints and quotas in the Amazon Web Services General Reference.

Valid Values: See Storage Gateway endpoints and quotas in the Amazon Web Services General Reference.

newActivateGateway

Arguments

:: Text
$sel:activationKey:ActivateGateway'

-> Text
$sel:gatewayName:ActivateGateway'

-> Text
$sel:gatewayTimezone:ActivateGateway'

-> Text
$sel:gatewayRegion:ActivateGateway'

-> ActivateGateway

Create a value of ActivateGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mediumChangerType:ActivateGateway', activateGateway_mediumChangerType - The value that indicates the type of medium changer to use for tape gateway. This field is optional.

Valid Values: STK-L700 | AWS-Gateway-VTL | IBM-03584L32-0402

$sel:tapeDriveType:ActivateGateway', activateGateway_tapeDriveType - The value that indicates the type of tape drive to use for tape gateway. This field is optional.

Valid Values: IBM-ULT3580-TD5

$sel:gatewayType:ActivateGateway', activateGateway_gatewayType - A value that defines the type of gateway to activate. The type specified is critical to all later functions of the gateway and cannot be changed after activation. The default value is CACHED.

Valid Values: STORED | CACHED | VTL | FILE_S3 | FILE_FSX_SMB|

$sel:tags:ActivateGateway', activateGateway_tags - A list of up to 50 tags that you can assign to the gateway. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers that can be represented in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256 characters.

$sel:activationKey:ActivateGateway', activateGateway_activationKey - Your gateway activation key. You can obtain the activation key by sending an HTTP GET request with redirects enabled to the gateway IP address (port 80). The redirect URL returned in the response provides you the activation key for your gateway in the query string parameter activationKey. It may also include other activation-related parameters, however, these are merely defaults -- the arguments you pass to the ActivateGateway API call determine the actual configuration of your gateway.

For more information, see Getting activation key in the Storage Gateway User Guide.

$sel:gatewayName:ActivateGateway', activateGateway_gatewayName - The name you configured for your gateway.

$sel:gatewayTimezone:ActivateGateway', activateGateway_gatewayTimezone - A value that indicates the time zone you want to set for the gateway. The time zone is of the format "GMT-hr:mm" or "GMT+hr:mm". For example, GMT-4:00 indicates the time is 4 hours behind GMT. GMT+2:00 indicates the time is 2 hours ahead of GMT. The time zone is used, for example, for scheduling snapshots and your gateway's maintenance schedule.

$sel:gatewayRegion:ActivateGateway', activateGateway_gatewayRegion - A value that indicates the Amazon Web Services Region where you want to store your data. The gateway Amazon Web Services Region specified must be the same Amazon Web Services Region as the Amazon Web Services Region in your Host header in the request. For more information about available Amazon Web Services Regions and endpoints for Storage Gateway, see Storage Gateway endpoints and quotas in the Amazon Web Services General Reference.

Valid Values: See Storage Gateway endpoints and quotas in the Amazon Web Services General Reference.

Request Lenses

activateGateway_mediumChangerType :: Lens' ActivateGateway (Maybe Text)

The value that indicates the type of medium changer to use for tape gateway. This field is optional.

Valid Values: STK-L700 | AWS-Gateway-VTL | IBM-03584L32-0402

activateGateway_tapeDriveType :: Lens' ActivateGateway (Maybe Text)

The value that indicates the type of tape drive to use for tape gateway. This field is optional.

Valid Values: IBM-ULT3580-TD5

activateGateway_gatewayType :: Lens' ActivateGateway (Maybe Text)

A value that defines the type of gateway to activate. The type specified is critical to all later functions of the gateway and cannot be changed after activation. The default value is CACHED.

Valid Values: STORED | CACHED | VTL | FILE_S3 | FILE_FSX_SMB|

activateGateway_tags :: Lens' ActivateGateway (Maybe [Tag])

A list of up to 50 tags that you can assign to the gateway. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers that can be represented in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256 characters.

activateGateway_activationKey :: Lens' ActivateGateway Text

Your gateway activation key. You can obtain the activation key by sending an HTTP GET request with redirects enabled to the gateway IP address (port 80). The redirect URL returned in the response provides you the activation key for your gateway in the query string parameter activationKey. It may also include other activation-related parameters, however, these are merely defaults -- the arguments you pass to the ActivateGateway API call determine the actual configuration of your gateway.

For more information, see Getting activation key in the Storage Gateway User Guide.

activateGateway_gatewayName :: Lens' ActivateGateway Text

The name you configured for your gateway.

activateGateway_gatewayTimezone :: Lens' ActivateGateway Text

A value that indicates the time zone you want to set for the gateway. The time zone is of the format "GMT-hr:mm" or "GMT+hr:mm". For example, GMT-4:00 indicates the time is 4 hours behind GMT. GMT+2:00 indicates the time is 2 hours ahead of GMT. The time zone is used, for example, for scheduling snapshots and your gateway's maintenance schedule.

activateGateway_gatewayRegion :: Lens' ActivateGateway Text

A value that indicates the Amazon Web Services Region where you want to store your data. The gateway Amazon Web Services Region specified must be the same Amazon Web Services Region as the Amazon Web Services Region in your Host header in the request. For more information about available Amazon Web Services Regions and endpoints for Storage Gateway, see Storage Gateway endpoints and quotas in the Amazon Web Services General Reference.

Valid Values: See Storage Gateway endpoints and quotas in the Amazon Web Services General Reference.

Destructuring the Response

data ActivateGatewayResponse

Storage Gateway returns the Amazon Resource Name (ARN) of the activated gateway. It is a string made of information such as your account, gateway name, and Amazon Web Services Region. This ARN is used to reference the gateway in other API operations as well as resource-based authorization.

For gateways activated prior to September 02, 2015, the gateway ARN contains the gateway name rather than the gateway ID. Changing the name of the gateway has no effect on the gateway ARN.

See: newActivateGatewayResponse smart constructor.

Constructors

ActivateGatewayResponse'

Fields

gatewayARN :: Maybe Text

httpStatus :: Int

The response's http status code.

newActivateGatewayResponse

Arguments

:: Int
$sel:httpStatus:ActivateGatewayResponse'

-> ActivateGatewayResponse

Create a value of ActivateGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:ActivateGatewayResponse', activateGatewayResponse_gatewayARN - Undocumented member.

$sel:httpStatus:ActivateGatewayResponse', activateGatewayResponse_httpStatus - The response's http status code.

Response Lenses

activateGatewayResponse_gatewayARN :: Lens' ActivateGatewayResponse (Maybe Text)

Undocumented member.

activateGatewayResponse_httpStatus :: Lens' ActivateGatewayResponse Int

The response's http status code.