Amazonka.StorageGateway.AddCache

Description

Configures one or more gateway local disks as cache for a gateway. This operation is only supported in the cached volume, tape, and file gateway type (see How Storage Gateway works (architecture).

In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add cache, and one or more disk IDs that you want to configure as cache.

Creating a Request

data AddCache

See: newAddCache smart constructor.

Constructors

AddCache'

Fields

gatewayARN :: Text

diskIds :: [Text]

An array of strings that identify disks that are to be configured as working storage. Each string has a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

newAddCache

Arguments

:: Text
$sel:gatewayARN:AddCache'

-> AddCache

Create a value of AddCache with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:AddCache', addCache_gatewayARN - Undocumented member.

$sel:diskIds:AddCache', addCache_diskIds - An array of strings that identify disks that are to be configured as working storage. Each string has a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

Request Lenses

addCache_gatewayARN :: Lens' AddCache Text

Undocumented member.

addCache_diskIds :: Lens' AddCache [Text]

An array of strings that identify disks that are to be configured as working storage. Each string has a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

Destructuring the Response

data AddCacheResponse

See: newAddCacheResponse smart constructor.

Constructors

AddCacheResponse'

Fields

gatewayARN :: Maybe Text

httpStatus :: Int

The response's http status code.

newAddCacheResponse

Arguments

:: Int
$sel:httpStatus:AddCacheResponse'

-> AddCacheResponse

Create a value of AddCacheResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:AddCache', addCacheResponse_gatewayARN - Undocumented member.

$sel:httpStatus:AddCacheResponse', addCacheResponse_httpStatus - The response's http status code.

Response Lenses

addCacheResponse_gatewayARN :: Lens' AddCacheResponse (Maybe Text)

Undocumented member.

addCacheResponse_httpStatus :: Lens' AddCacheResponse Int

The response's http status code.