Amazonka.StorageGateway.AssociateFileSystem

Description

Associate an Amazon FSx file system with the FSx File Gateway. After the association process is complete, the file shares on the Amazon FSx file system are available for access through the gateway. This operation only supports the FSx File Gateway type.

Creating a Request

data AssociateFileSystem

See: newAssociateFileSystem smart constructor.

Constructors

AssociateFileSystem'

Fields

auditDestinationARN :: Maybe Text

The Amazon Resource Name (ARN) of the storage used for the audit logs.

cacheAttributes :: Maybe CacheAttributes

endpointNetworkConfiguration :: Maybe EndpointNetworkConfiguration

Specifies the network configuration information for the gateway associated with the Amazon FSx file system.

If multiple file systems are associated with this gateway, this parameter's IpAddresses field is required.

tags :: Maybe [Tag]

A list of up to 50 tags that can be assigned to the file system association. Each tag is a key-value pair.

userName :: Text

The user name of the user credential that has permission to access the root share D$ of the Amazon FSx file system. The user account must belong to the Amazon FSx delegated admin user group.

password :: Sensitive Text

The password of the user credential.

clientToken :: Text

A unique string value that you supply that is used by the FSx File Gateway to ensure idempotent file system association creation.

gatewayARN :: Text

locationARN :: Text

The Amazon Resource Name (ARN) of the Amazon FSx file system to associate with the FSx File Gateway.

newAssociateFileSystem

Arguments

:: Text
$sel:userName:AssociateFileSystem'

-> Text
$sel:password:AssociateFileSystem'

-> Text
$sel:clientToken:AssociateFileSystem'

-> Text
$sel:gatewayARN:AssociateFileSystem'

-> Text
$sel:locationARN:AssociateFileSystem'

-> AssociateFileSystem

Create a value of AssociateFileSystem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:auditDestinationARN:AssociateFileSystem', associateFileSystem_auditDestinationARN - The Amazon Resource Name (ARN) of the storage used for the audit logs.

$sel:cacheAttributes:AssociateFileSystem', associateFileSystem_cacheAttributes - Undocumented member.

$sel:endpointNetworkConfiguration:AssociateFileSystem', associateFileSystem_endpointNetworkConfiguration - Specifies the network configuration information for the gateway associated with the Amazon FSx file system.

If multiple file systems are associated with this gateway, this parameter's IpAddresses field is required.

$sel:tags:AssociateFileSystem', associateFileSystem_tags - A list of up to 50 tags that can be assigned to the file system association. Each tag is a key-value pair.

$sel:userName:AssociateFileSystem', associateFileSystem_userName - The user name of the user credential that has permission to access the root share D$ of the Amazon FSx file system. The user account must belong to the Amazon FSx delegated admin user group.

$sel:password:AssociateFileSystem', associateFileSystem_password - The password of the user credential.

$sel:clientToken:AssociateFileSystem', associateFileSystem_clientToken - A unique string value that you supply that is used by the FSx File Gateway to ensure idempotent file system association creation.

$sel:gatewayARN:AssociateFileSystem', associateFileSystem_gatewayARN - Undocumented member.

$sel:locationARN:AssociateFileSystem', associateFileSystem_locationARN - The Amazon Resource Name (ARN) of the Amazon FSx file system to associate with the FSx File Gateway.

Request Lenses

associateFileSystem_auditDestinationARN :: Lens' AssociateFileSystem (Maybe Text)

The Amazon Resource Name (ARN) of the storage used for the audit logs.

associateFileSystem_cacheAttributes :: Lens' AssociateFileSystem (Maybe CacheAttributes)

Undocumented member.

associateFileSystem_endpointNetworkConfiguration :: Lens' AssociateFileSystem (Maybe EndpointNetworkConfiguration)

Specifies the network configuration information for the gateway associated with the Amazon FSx file system.

If multiple file systems are associated with this gateway, this parameter's IpAddresses field is required.

associateFileSystem_tags :: Lens' AssociateFileSystem (Maybe [Tag])

A list of up to 50 tags that can be assigned to the file system association. Each tag is a key-value pair.

associateFileSystem_userName :: Lens' AssociateFileSystem Text

The user name of the user credential that has permission to access the root share D$ of the Amazon FSx file system. The user account must belong to the Amazon FSx delegated admin user group.

associateFileSystem_password :: Lens' AssociateFileSystem Text

The password of the user credential.

associateFileSystem_clientToken :: Lens' AssociateFileSystem Text

A unique string value that you supply that is used by the FSx File Gateway to ensure idempotent file system association creation.

associateFileSystem_gatewayARN :: Lens' AssociateFileSystem Text

Undocumented member.

associateFileSystem_locationARN :: Lens' AssociateFileSystem Text

The Amazon Resource Name (ARN) of the Amazon FSx file system to associate with the FSx File Gateway.

Destructuring the Response

data AssociateFileSystemResponse

See: newAssociateFileSystemResponse smart constructor.

Constructors

AssociateFileSystemResponse'

Fields

fileSystemAssociationARN :: Maybe Text

The ARN of the newly created file system association.

httpStatus :: Int

The response's http status code.

newAssociateFileSystemResponse

Arguments

:: Int
$sel:httpStatus:AssociateFileSystemResponse'

-> AssociateFileSystemResponse

Create a value of AssociateFileSystemResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileSystemAssociationARN:AssociateFileSystemResponse', associateFileSystemResponse_fileSystemAssociationARN - The ARN of the newly created file system association.

$sel:httpStatus:AssociateFileSystemResponse', associateFileSystemResponse_httpStatus - The response's http status code.

Response Lenses

associateFileSystemResponse_fileSystemAssociationARN :: Lens' AssociateFileSystemResponse (Maybe Text)

The ARN of the newly created file system association.

associateFileSystemResponse_httpStatus :: Lens' AssociateFileSystemResponse Int

The response's http status code.