Amazonka.StorageGateway.CreateCachediSCSIVolume

Description

Creates a cached volume on a specified cached volume gateway. This operation is only supported in the cached volume gateway type.

Cache storage must be allocated to the gateway before you can create a cached volume. Use the AddCache operation to add cache storage to a gateway.

In the request, you must specify the gateway, size of the volume in bytes, the iSCSI target name, an IP address on which to expose the target, and a unique client token. In response, the gateway creates the volume and returns information about it. This information includes the volume Amazon Resource Name (ARN), its size, and the iSCSI target ARN that initiators can use to connect to the volume target.

Optionally, you can provide the ARN for an existing volume as the SourceVolumeARN for this cached volume, which creates an exact copy of the existing volume’s latest recovery point. The VolumeSizeInBytes value must be equal to or larger than the size of the copied volume, in bytes.

Creating a Request

data CreateCachediSCSIVolume

See: newCreateCachediSCSIVolume smart constructor.

Constructors

CreateCachediSCSIVolume'

Fields

kmsKey :: Maybe Text

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

sourceVolumeARN :: Maybe Text

The ARN for an existing volume. Specifying this ARN makes the new volume into an exact copy of the specified existing volume's latest recovery point. The VolumeSizeInBytes value for this new volume must be equal to or larger than the size of the existing volume, in bytes.

kmsEncrypted :: Maybe Bool

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

tags :: Maybe [Tag]

A list of up to 50 tags that you can assign to a cached volume. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers that you can represent in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256 characters.

snapshotId :: Maybe Text

The snapshot ID (e.g. "snap-1122aabb") of the snapshot to restore as the new cached volume. Specify this field if you want to create the iSCSI storage volume from a snapshot; otherwise, do not include this field. To list snapshots for your account use DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference.

gatewayARN :: Text

volumeSizeInBytes :: Integer

The size of the volume in bytes.

targetName :: Text

The name of the iSCSI target used by an initiator to connect to a volume and used as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes on a gateway.

If you don't specify a value, Storage Gateway uses the value that was previously used for this volume as the new target name.

networkInterfaceId :: Text

The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway.

Valid Values: A valid IP address.

clientToken :: Text

A unique identifier that you use to retry a request. If you retry a request, use the same ClientToken you specified in the initial request.

newCreateCachediSCSIVolume

Arguments

:: Text
$sel:gatewayARN:CreateCachediSCSIVolume'

-> Integer
$sel:volumeSizeInBytes:CreateCachediSCSIVolume'

-> Text
$sel:targetName:CreateCachediSCSIVolume'

-> Text
$sel:networkInterfaceId:CreateCachediSCSIVolume'

-> Text
$sel:clientToken:CreateCachediSCSIVolume'

-> CreateCachediSCSIVolume

Create a value of CreateCachediSCSIVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKey:CreateCachediSCSIVolume', createCachediSCSIVolume_kmsKey - The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

$sel:sourceVolumeARN:CreateCachediSCSIVolume', createCachediSCSIVolume_sourceVolumeARN - The ARN for an existing volume. Specifying this ARN makes the new volume into an exact copy of the specified existing volume's latest recovery point. The VolumeSizeInBytes value for this new volume must be equal to or larger than the size of the existing volume, in bytes.

$sel:kmsEncrypted:CreateCachediSCSIVolume', createCachediSCSIVolume_kmsEncrypted - Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

$sel:tags:CreateCachediSCSIVolume', createCachediSCSIVolume_tags - A list of up to 50 tags that you can assign to a cached volume. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers that you can represent in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256 characters.

$sel:snapshotId:CreateCachediSCSIVolume', createCachediSCSIVolume_snapshotId - The snapshot ID (e.g. "snap-1122aabb") of the snapshot to restore as the new cached volume. Specify this field if you want to create the iSCSI storage volume from a snapshot; otherwise, do not include this field. To list snapshots for your account use DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference.

$sel:gatewayARN:CreateCachediSCSIVolume', createCachediSCSIVolume_gatewayARN - Undocumented member.

$sel:volumeSizeInBytes:CreateCachediSCSIVolume', createCachediSCSIVolume_volumeSizeInBytes - The size of the volume in bytes.

$sel:targetName:CreateCachediSCSIVolume', createCachediSCSIVolume_targetName - The name of the iSCSI target used by an initiator to connect to a volume and used as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes on a gateway.

If you don't specify a value, Storage Gateway uses the value that was previously used for this volume as the new target name.

$sel:networkInterfaceId:CreateCachediSCSIVolume', createCachediSCSIVolume_networkInterfaceId - The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway.

Valid Values: A valid IP address.

$sel:clientToken:CreateCachediSCSIVolume', createCachediSCSIVolume_clientToken - A unique identifier that you use to retry a request. If you retry a request, use the same ClientToken you specified in the initial request.

Request Lenses

createCachediSCSIVolume_kmsKey :: Lens' CreateCachediSCSIVolume (Maybe Text)

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

createCachediSCSIVolume_sourceVolumeARN :: Lens' CreateCachediSCSIVolume (Maybe Text)

The ARN for an existing volume. Specifying this ARN makes the new volume into an exact copy of the specified existing volume's latest recovery point. The VolumeSizeInBytes value for this new volume must be equal to or larger than the size of the existing volume, in bytes.

createCachediSCSIVolume_kmsEncrypted :: Lens' CreateCachediSCSIVolume (Maybe Bool)

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

createCachediSCSIVolume_tags :: Lens' CreateCachediSCSIVolume (Maybe [Tag])

A list of up to 50 tags that you can assign to a cached volume. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers that you can represent in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256 characters.

createCachediSCSIVolume_snapshotId :: Lens' CreateCachediSCSIVolume (Maybe Text)

The snapshot ID (e.g. "snap-1122aabb") of the snapshot to restore as the new cached volume. Specify this field if you want to create the iSCSI storage volume from a snapshot; otherwise, do not include this field. To list snapshots for your account use DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference.

createCachediSCSIVolume_gatewayARN :: Lens' CreateCachediSCSIVolume Text

Undocumented member.

createCachediSCSIVolume_volumeSizeInBytes :: Lens' CreateCachediSCSIVolume Integer

The size of the volume in bytes.

createCachediSCSIVolume_targetName :: Lens' CreateCachediSCSIVolume Text

The name of the iSCSI target used by an initiator to connect to a volume and used as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes on a gateway.

If you don't specify a value, Storage Gateway uses the value that was previously used for this volume as the new target name.

createCachediSCSIVolume_networkInterfaceId :: Lens' CreateCachediSCSIVolume Text

The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway.

Valid Values: A valid IP address.

createCachediSCSIVolume_clientToken :: Lens' CreateCachediSCSIVolume Text

A unique identifier that you use to retry a request. If you retry a request, use the same ClientToken you specified in the initial request.

Destructuring the Response

data CreateCachediSCSIVolumeResponse

See: newCreateCachediSCSIVolumeResponse smart constructor.

Constructors

CreateCachediSCSIVolumeResponse'

Fields

targetARN :: Maybe Text

The Amazon Resource Name (ARN) of the volume target, which includes the iSCSI name that initiators can use to connect to the target.

volumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the configured volume.

httpStatus :: Int

The response's http status code.

newCreateCachediSCSIVolumeResponse

Arguments

:: Int
$sel:httpStatus:CreateCachediSCSIVolumeResponse'

-> CreateCachediSCSIVolumeResponse

Create a value of CreateCachediSCSIVolumeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetARN:CreateCachediSCSIVolumeResponse', createCachediSCSIVolumeResponse_targetARN - The Amazon Resource Name (ARN) of the volume target, which includes the iSCSI name that initiators can use to connect to the target.

$sel:volumeARN:CreateCachediSCSIVolumeResponse', createCachediSCSIVolumeResponse_volumeARN - The Amazon Resource Name (ARN) of the configured volume.

$sel:httpStatus:CreateCachediSCSIVolumeResponse', createCachediSCSIVolumeResponse_httpStatus - The response's http status code.

Response Lenses

createCachediSCSIVolumeResponse_targetARN :: Lens' CreateCachediSCSIVolumeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the volume target, which includes the iSCSI name that initiators can use to connect to the target.

createCachediSCSIVolumeResponse_volumeARN :: Lens' CreateCachediSCSIVolumeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the configured volume.

createCachediSCSIVolumeResponse_httpStatus :: Lens' CreateCachediSCSIVolumeResponse Int

The response's http status code.