Amazonka.StorageGateway.CreateNFSFileShare

Description

Creates a Network File System (NFS) file share on an existing S3 File Gateway. In Storage Gateway, a file share is a file system mount point backed by Amazon S3 cloud storage. Storage Gateway exposes file shares using an NFS interface. This operation is only supported for S3 File Gateways.

S3 File gateway requires Security Token Service (Amazon Web Services STS) to be activated to enable you to create a file share. Make sure Amazon Web Services STS is activated in the Amazon Web Services Region you are creating your S3 File Gateway in. If Amazon Web Services STS is not activated in the Amazon Web Services Region, activate it. For information about how to activate Amazon Web Services STS, see Activating and deactivating Amazon Web Services STS in an Amazon Web Services Region in the Identity and Access Management User Guide.

S3 File Gateways do not support creating hard or symbolic links on a file share.

Creating a Request

data CreateNFSFileShare

CreateNFSFileShareInput

See: newCreateNFSFileShare smart constructor.

Constructors

CreateNFSFileShare'

Fields

auditDestinationARN :: Maybe Text

The Amazon Resource Name (ARN) of the storage used for audit logs.

kmsKey :: Maybe Text

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

vPCEndpointDNSName :: Maybe Text

Specifies the DNS name for the VPC endpoint that the NFS file share uses to connect to Amazon S3.

This parameter is required for NFS file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

cacheAttributes :: Maybe CacheAttributes

Specifies refresh cache information for the file share.

objectACL :: Maybe ObjectACL

A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

kmsEncrypted :: Maybe Bool

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

defaultStorageClass :: Maybe Text

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

fileShareName :: Maybe Text

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN, or if an access point or access point alias is used.

notificationPolicy :: Maybe Text

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

squash :: Maybe Text

A value that maps a user to anonymous user.

Valid values are the following:

RootSquash: Only root is mapped to anonymous user.

NoSquash: No one is mapped to anonymous user.

AllSquash: Everyone is mapped to anonymous user.

requesterPays :: Maybe Bool

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

nFSFileShareDefaults :: Maybe NFSFileShareDefaults

File share default values. Optional.

clientList :: Maybe (NonEmpty Text)

The list of clients that are allowed to access the S3 File Gateway. The list must contain either valid IP addresses or valid CIDR blocks.

guessMIMETypeEnabled :: Maybe Bool

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

readOnly :: Maybe Bool

A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

bucketRegion :: Maybe Text

Specifies the Region of the S3 bucket where the NFS file share stores files.

This parameter is required for NFS file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

tags :: Maybe [Tag]

A list of up to 50 tags that can be assigned to the NFS file share. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

clientToken :: Text

A unique string value that you supply that is used by S3 File Gateway to ensure idempotent file share creation.

gatewayARN :: Text

The Amazon Resource Name (ARN) of the S3 File Gateway on which you want to create a file share.

role' :: Text

The ARN of the Identity and Access Management (IAM) role that an S3 File Gateway assumes when it accesses the underlying storage.

locationARN :: Text

A custom ARN for the backend storage used for storing data for file shares. It includes a resource ARN with an optional prefix concatenation. The prefix must end with a forward slash (/).

You can specify LocationARN as a bucket ARN, access point ARN or access point alias, as shown in the following examples.

Bucket ARN:

arn:aws:s3:::my-bucket/prefix/

Access point ARN:

arn:aws:s3:region:account-id:accesspoint/access-point-name/prefix/

If you specify an access point, the bucket policy must be configured to delegate access control to the access point. For information, see Delegating access control to access points in the Amazon S3 User Guide.

Access point alias:

test-ap-ab123cdef4gehijklmn5opqrstuvuse1a-s3alias

newCreateNFSFileShare

Arguments

:: Text
$sel:clientToken:CreateNFSFileShare'

-> Text
$sel:gatewayARN:CreateNFSFileShare'

-> Text
$sel:role':CreateNFSFileShare'

-> Text
$sel:locationARN:CreateNFSFileShare'

-> CreateNFSFileShare

Create a value of CreateNFSFileShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:auditDestinationARN:CreateNFSFileShare', createNFSFileShare_auditDestinationARN - The Amazon Resource Name (ARN) of the storage used for audit logs.

$sel:kmsKey:CreateNFSFileShare', createNFSFileShare_kmsKey - The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

$sel:vPCEndpointDNSName:CreateNFSFileShare', createNFSFileShare_vPCEndpointDNSName - Specifies the DNS name for the VPC endpoint that the NFS file share uses to connect to Amazon S3.

This parameter is required for NFS file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

$sel:cacheAttributes:CreateNFSFileShare', createNFSFileShare_cacheAttributes - Specifies refresh cache information for the file share.

$sel:objectACL:CreateNFSFileShare', createNFSFileShare_objectACL - A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

$sel:kmsEncrypted:CreateNFSFileShare', createNFSFileShare_kmsEncrypted - Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

$sel:defaultStorageClass:CreateNFSFileShare', createNFSFileShare_defaultStorageClass - The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

$sel:fileShareName:CreateNFSFileShare', createNFSFileShare_fileShareName - The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN, or if an access point or access point alias is used.

$sel:notificationPolicy:CreateNFSFileShare', createNFSFileShare_notificationPolicy - The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

$sel:squash:CreateNFSFileShare', createNFSFileShare_squash - A value that maps a user to anonymous user.

Valid values are the following:

RootSquash: Only root is mapped to anonymous user.

NoSquash: No one is mapped to anonymous user.

AllSquash: Everyone is mapped to anonymous user.

$sel:requesterPays:CreateNFSFileShare', createNFSFileShare_requesterPays - A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

$sel:nFSFileShareDefaults:CreateNFSFileShare', createNFSFileShare_nFSFileShareDefaults - File share default values. Optional.

$sel:clientList:CreateNFSFileShare', createNFSFileShare_clientList - The list of clients that are allowed to access the S3 File Gateway. The list must contain either valid IP addresses or valid CIDR blocks.

$sel:guessMIMETypeEnabled:CreateNFSFileShare', createNFSFileShare_guessMIMETypeEnabled - A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

$sel:readOnly:CreateNFSFileShare', createNFSFileShare_readOnly - A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

$sel:bucketRegion:CreateNFSFileShare', createNFSFileShare_bucketRegion - Specifies the Region of the S3 bucket where the NFS file share stores files.

This parameter is required for NFS file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

$sel:tags:CreateNFSFileShare', createNFSFileShare_tags - A list of up to 50 tags that can be assigned to the NFS file share. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

$sel:clientToken:CreateNFSFileShare', createNFSFileShare_clientToken - A unique string value that you supply that is used by S3 File Gateway to ensure idempotent file share creation.

$sel:gatewayARN:CreateNFSFileShare', createNFSFileShare_gatewayARN - The Amazon Resource Name (ARN) of the S3 File Gateway on which you want to create a file share.

$sel:role':CreateNFSFileShare', createNFSFileShare_role - The ARN of the Identity and Access Management (IAM) role that an S3 File Gateway assumes when it accesses the underlying storage.

$sel:locationARN:CreateNFSFileShare', createNFSFileShare_locationARN - A custom ARN for the backend storage used for storing data for file shares. It includes a resource ARN with an optional prefix concatenation. The prefix must end with a forward slash (/).

You can specify LocationARN as a bucket ARN, access point ARN or access point alias, as shown in the following examples.

Bucket ARN:

arn:aws:s3:::my-bucket/prefix/

Access point ARN:

arn:aws:s3:region:account-id:accesspoint/access-point-name/prefix/

If you specify an access point, the bucket policy must be configured to delegate access control to the access point. For information, see Delegating access control to access points in the Amazon S3 User Guide.

Access point alias:

test-ap-ab123cdef4gehijklmn5opqrstuvuse1a-s3alias

Request Lenses

createNFSFileShare_auditDestinationARN :: Lens' CreateNFSFileShare (Maybe Text)

The Amazon Resource Name (ARN) of the storage used for audit logs.

createNFSFileShare_kmsKey :: Lens' CreateNFSFileShare (Maybe Text)

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

createNFSFileShare_vPCEndpointDNSName :: Lens' CreateNFSFileShare (Maybe Text)

Specifies the DNS name for the VPC endpoint that the NFS file share uses to connect to Amazon S3.

This parameter is required for NFS file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

createNFSFileShare_cacheAttributes :: Lens' CreateNFSFileShare (Maybe CacheAttributes)

Specifies refresh cache information for the file share.

createNFSFileShare_objectACL :: Lens' CreateNFSFileShare (Maybe ObjectACL)

A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

createNFSFileShare_kmsEncrypted :: Lens' CreateNFSFileShare (Maybe Bool)

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

createNFSFileShare_defaultStorageClass :: Lens' CreateNFSFileShare (Maybe Text)

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

createNFSFileShare_fileShareName :: Lens' CreateNFSFileShare (Maybe Text)

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN, or if an access point or access point alias is used.

createNFSFileShare_notificationPolicy :: Lens' CreateNFSFileShare (Maybe Text)

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

createNFSFileShare_squash :: Lens' CreateNFSFileShare (Maybe Text)

A value that maps a user to anonymous user.

Valid values are the following:

RootSquash: Only root is mapped to anonymous user.

NoSquash: No one is mapped to anonymous user.

AllSquash: Everyone is mapped to anonymous user.

createNFSFileShare_requesterPays :: Lens' CreateNFSFileShare (Maybe Bool)

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

createNFSFileShare_nFSFileShareDefaults :: Lens' CreateNFSFileShare (Maybe NFSFileShareDefaults)

File share default values. Optional.

createNFSFileShare_clientList :: Lens' CreateNFSFileShare (Maybe (NonEmpty Text))

The list of clients that are allowed to access the S3 File Gateway. The list must contain either valid IP addresses or valid CIDR blocks.

createNFSFileShare_guessMIMETypeEnabled :: Lens' CreateNFSFileShare (Maybe Bool)

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

createNFSFileShare_readOnly :: Lens' CreateNFSFileShare (Maybe Bool)

A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

createNFSFileShare_bucketRegion :: Lens' CreateNFSFileShare (Maybe Text)

Specifies the Region of the S3 bucket where the NFS file share stores files.

This parameter is required for NFS file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

createNFSFileShare_tags :: Lens' CreateNFSFileShare (Maybe [Tag])

A list of up to 50 tags that can be assigned to the NFS file share. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

createNFSFileShare_clientToken :: Lens' CreateNFSFileShare Text

A unique string value that you supply that is used by S3 File Gateway to ensure idempotent file share creation.

createNFSFileShare_gatewayARN :: Lens' CreateNFSFileShare Text

The Amazon Resource Name (ARN) of the S3 File Gateway on which you want to create a file share.

createNFSFileShare_role :: Lens' CreateNFSFileShare Text

The ARN of the Identity and Access Management (IAM) role that an S3 File Gateway assumes when it accesses the underlying storage.

createNFSFileShare_locationARN :: Lens' CreateNFSFileShare Text

A custom ARN for the backend storage used for storing data for file shares. It includes a resource ARN with an optional prefix concatenation. The prefix must end with a forward slash (/).

You can specify LocationARN as a bucket ARN, access point ARN or access point alias, as shown in the following examples.

Bucket ARN:

arn:aws:s3:::my-bucket/prefix/

Access point ARN:

arn:aws:s3:region:account-id:accesspoint/access-point-name/prefix/

If you specify an access point, the bucket policy must be configured to delegate access control to the access point. For information, see Delegating access control to access points in the Amazon S3 User Guide.

Access point alias:

test-ap-ab123cdef4gehijklmn5opqrstuvuse1a-s3alias

Destructuring the Response

data CreateNFSFileShareResponse

CreateNFSFileShareOutput

See: newCreateNFSFileShareResponse smart constructor.

Constructors

CreateNFSFileShareResponse'

Fields

fileShareARN :: Maybe Text

The Amazon Resource Name (ARN) of the newly created file share.

httpStatus :: Int

The response's http status code.

newCreateNFSFileShareResponse

Arguments

:: Int
$sel:httpStatus:CreateNFSFileShareResponse'

-> CreateNFSFileShareResponse

Create a value of CreateNFSFileShareResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileShareARN:CreateNFSFileShareResponse', createNFSFileShareResponse_fileShareARN - The Amazon Resource Name (ARN) of the newly created file share.

$sel:httpStatus:CreateNFSFileShareResponse', createNFSFileShareResponse_httpStatus - The response's http status code.

Response Lenses

createNFSFileShareResponse_fileShareARN :: Lens' CreateNFSFileShareResponse (Maybe Text)

The Amazon Resource Name (ARN) of the newly created file share.

createNFSFileShareResponse_httpStatus :: Lens' CreateNFSFileShareResponse Int

The response's http status code.