Amazonka.StorageGateway.CreateSMBFileShare

Description

Creates a Server Message Block (SMB) file share on an existing S3 File Gateway. In Storage Gateway, a file share is a file system mount point backed by Amazon S3 cloud storage. Storage Gateway exposes file shares using an SMB interface. This operation is only supported for S3 File Gateways.

S3 File Gateways require Security Token Service (Amazon Web Services STS) to be activated to enable you to create a file share. Make sure that Amazon Web Services STS is activated in the Amazon Web Services Region you are creating your S3 File Gateway in. If Amazon Web Services STS is not activated in this Amazon Web Services Region, activate it. For information about how to activate Amazon Web Services STS, see Activating and deactivating Amazon Web Services STS in an Amazon Web Services Region in the Identity and Access Management User Guide.

File gateways don't support creating hard or symbolic links on a file share.

Creating a Request

data CreateSMBFileShare

CreateSMBFileShareInput

See: newCreateSMBFileShare smart constructor.

Constructors

CreateSMBFileShare'

Fields

accessBasedEnumeration :: Maybe Bool

The files and folders on this share will only be visible to users with read access.

adminUserList :: Maybe [Text]

A list of users or groups in the Active Directory that will be granted administrator privileges on the file share. These users can do all file operations as the super-user. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1.

Use this option very carefully, because any user in this list can do anything they like on the file share, regardless of file permissions.

auditDestinationARN :: Maybe Text

The Amazon Resource Name (ARN) of the storage used for audit logs.

invalidUserList :: Maybe [Text]

A list of users or groups in the Active Directory that are not allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

kmsKey :: Maybe Text

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

validUserList :: Maybe [Text]

A list of users or groups in the Active Directory that are allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

vPCEndpointDNSName :: Maybe Text

Specifies the DNS name for the VPC endpoint that the SMB file share uses to connect to Amazon S3.

This parameter is required for SMB file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

authentication :: Maybe Text

The authentication method that users use to access the file share. The default is ActiveDirectory.

Valid Values: ActiveDirectory | GuestAccess

cacheAttributes :: Maybe CacheAttributes

Specifies refresh cache information for the file share.

objectACL :: Maybe ObjectACL

A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

kmsEncrypted :: Maybe Bool

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

defaultStorageClass :: Maybe Text

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

fileShareName :: Maybe Text

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN, or if an access point or access point alias is used.

sMBACLEnabled :: Maybe Bool

Set this value to true to enable access control list (ACL) on the SMB file share. Set it to false to map file and directory permissions to the POSIX permissions.

For more information, see Using Microsoft Windows ACLs to control access to an SMB file share in the Storage Gateway User Guide.

Valid Values: true | false

oplocksEnabled :: Maybe Bool

Specifies whether opportunistic locking is enabled for the SMB file share.

Enabling opportunistic locking on case-sensitive shares is not recommended for workloads that involve access to files with the same name in different case.

Valid Values: true | false

notificationPolicy :: Maybe Text

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

requesterPays :: Maybe Bool

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

guessMIMETypeEnabled :: Maybe Bool

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

readOnly :: Maybe Bool

A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

bucketRegion :: Maybe Text

Specifies the Region of the S3 bucket where the SMB file share stores files.

This parameter is required for SMB file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

caseSensitivity :: Maybe CaseSensitivity

The case of an object name in an Amazon S3 bucket. For ClientSpecified, the client determines the case sensitivity. For CaseSensitive, the gateway determines the case sensitivity. The default value is ClientSpecified.

tags :: Maybe [Tag]

A list of up to 50 tags that can be assigned to the NFS file share. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

clientToken :: Text

A unique string value that you supply that is used by S3 File Gateway to ensure idempotent file share creation.

gatewayARN :: Text

The ARN of the S3 File Gateway on which you want to create a file share.

role' :: Text

The ARN of the Identity and Access Management (IAM) role that an S3 File Gateway assumes when it accesses the underlying storage.

locationARN :: Text

A custom ARN for the backend storage used for storing data for file shares. It includes a resource ARN with an optional prefix concatenation. The prefix must end with a forward slash (/).

You can specify LocationARN as a bucket ARN, access point ARN or access point alias, as shown in the following examples.

Bucket ARN:

arn:aws:s3:::my-bucket/prefix/

Access point ARN:

arn:aws:s3:region:account-id:accesspoint/access-point-name/prefix/

If you specify an access point, the bucket policy must be configured to delegate access control to the access point. For information, see Delegating access control to access points in the Amazon S3 User Guide.

Access point alias:

test-ap-ab123cdef4gehijklmn5opqrstuvuse1a-s3alias

newCreateSMBFileShare

Arguments

:: Text
$sel:clientToken:CreateSMBFileShare'

-> Text
$sel:gatewayARN:CreateSMBFileShare'

-> Text
$sel:role':CreateSMBFileShare'

-> Text
$sel:locationARN:CreateSMBFileShare'

-> CreateSMBFileShare

Create a value of CreateSMBFileShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessBasedEnumeration:CreateSMBFileShare', createSMBFileShare_accessBasedEnumeration - The files and folders on this share will only be visible to users with read access.

$sel:adminUserList:CreateSMBFileShare', createSMBFileShare_adminUserList - A list of users or groups in the Active Directory that will be granted administrator privileges on the file share. These users can do all file operations as the super-user. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1.

Use this option very carefully, because any user in this list can do anything they like on the file share, regardless of file permissions.

$sel:auditDestinationARN:CreateSMBFileShare', createSMBFileShare_auditDestinationARN - The Amazon Resource Name (ARN) of the storage used for audit logs.

$sel:invalidUserList:CreateSMBFileShare', createSMBFileShare_invalidUserList - A list of users or groups in the Active Directory that are not allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

$sel:kmsKey:CreateSMBFileShare', createSMBFileShare_kmsKey - The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

$sel:validUserList:CreateSMBFileShare', createSMBFileShare_validUserList - A list of users or groups in the Active Directory that are allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

$sel:vPCEndpointDNSName:CreateSMBFileShare', createSMBFileShare_vPCEndpointDNSName - Specifies the DNS name for the VPC endpoint that the SMB file share uses to connect to Amazon S3.

This parameter is required for SMB file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

$sel:authentication:CreateSMBFileShare', createSMBFileShare_authentication - The authentication method that users use to access the file share. The default is ActiveDirectory.

Valid Values: ActiveDirectory | GuestAccess

$sel:cacheAttributes:CreateSMBFileShare', createSMBFileShare_cacheAttributes - Specifies refresh cache information for the file share.

$sel:objectACL:CreateSMBFileShare', createSMBFileShare_objectACL - A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

$sel:kmsEncrypted:CreateSMBFileShare', createSMBFileShare_kmsEncrypted - Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

$sel:defaultStorageClass:CreateSMBFileShare', createSMBFileShare_defaultStorageClass - The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

$sel:fileShareName:CreateSMBFileShare', createSMBFileShare_fileShareName - The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN, or if an access point or access point alias is used.

$sel:sMBACLEnabled:CreateSMBFileShare', createSMBFileShare_sMBACLEnabled - Set this value to true to enable access control list (ACL) on the SMB file share. Set it to false to map file and directory permissions to the POSIX permissions.

For more information, see Using Microsoft Windows ACLs to control access to an SMB file share in the Storage Gateway User Guide.

Valid Values: true | false

$sel:oplocksEnabled:CreateSMBFileShare', createSMBFileShare_oplocksEnabled - Specifies whether opportunistic locking is enabled for the SMB file share.

Enabling opportunistic locking on case-sensitive shares is not recommended for workloads that involve access to files with the same name in different case.

Valid Values: true | false

$sel:notificationPolicy:CreateSMBFileShare', createSMBFileShare_notificationPolicy - The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

$sel:requesterPays:CreateSMBFileShare', createSMBFileShare_requesterPays - A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

$sel:guessMIMETypeEnabled:CreateSMBFileShare', createSMBFileShare_guessMIMETypeEnabled - A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

$sel:readOnly:CreateSMBFileShare', createSMBFileShare_readOnly - A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

$sel:bucketRegion:CreateSMBFileShare', createSMBFileShare_bucketRegion - Specifies the Region of the S3 bucket where the SMB file share stores files.

This parameter is required for SMB file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

$sel:caseSensitivity:CreateSMBFileShare', createSMBFileShare_caseSensitivity - The case of an object name in an Amazon S3 bucket. For ClientSpecified, the client determines the case sensitivity. For CaseSensitive, the gateway determines the case sensitivity. The default value is ClientSpecified.

$sel:tags:CreateSMBFileShare', createSMBFileShare_tags - A list of up to 50 tags that can be assigned to the NFS file share. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

$sel:clientToken:CreateSMBFileShare', createSMBFileShare_clientToken - A unique string value that you supply that is used by S3 File Gateway to ensure idempotent file share creation.

$sel:gatewayARN:CreateSMBFileShare', createSMBFileShare_gatewayARN - The ARN of the S3 File Gateway on which you want to create a file share.

$sel:role':CreateSMBFileShare', createSMBFileShare_role - The ARN of the Identity and Access Management (IAM) role that an S3 File Gateway assumes when it accesses the underlying storage.

$sel:locationARN:CreateSMBFileShare', createSMBFileShare_locationARN - A custom ARN for the backend storage used for storing data for file shares. It includes a resource ARN with an optional prefix concatenation. The prefix must end with a forward slash (/).

You can specify LocationARN as a bucket ARN, access point ARN or access point alias, as shown in the following examples.

Bucket ARN:

arn:aws:s3:::my-bucket/prefix/

Access point ARN:

arn:aws:s3:region:account-id:accesspoint/access-point-name/prefix/

If you specify an access point, the bucket policy must be configured to delegate access control to the access point. For information, see Delegating access control to access points in the Amazon S3 User Guide.

Access point alias:

test-ap-ab123cdef4gehijklmn5opqrstuvuse1a-s3alias

Request Lenses

createSMBFileShare_accessBasedEnumeration :: Lens' CreateSMBFileShare (Maybe Bool)

The files and folders on this share will only be visible to users with read access.

createSMBFileShare_adminUserList :: Lens' CreateSMBFileShare (Maybe [Text])

A list of users or groups in the Active Directory that will be granted administrator privileges on the file share. These users can do all file operations as the super-user. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1.

Use this option very carefully, because any user in this list can do anything they like on the file share, regardless of file permissions.

createSMBFileShare_auditDestinationARN :: Lens' CreateSMBFileShare (Maybe Text)

The Amazon Resource Name (ARN) of the storage used for audit logs.

createSMBFileShare_invalidUserList :: Lens' CreateSMBFileShare (Maybe [Text])

A list of users or groups in the Active Directory that are not allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

createSMBFileShare_kmsKey :: Lens' CreateSMBFileShare (Maybe Text)

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

createSMBFileShare_validUserList :: Lens' CreateSMBFileShare (Maybe [Text])

A list of users or groups in the Active Directory that are allowed to access the file share. A group must be prefixed with the @ character. Acceptable formats include: DOMAIN\User1, user1, @group1, and @DOMAIN\group1. Can only be set if Authentication is set to ActiveDirectory.

createSMBFileShare_vPCEndpointDNSName :: Lens' CreateSMBFileShare (Maybe Text)

Specifies the DNS name for the VPC endpoint that the SMB file share uses to connect to Amazon S3.

This parameter is required for SMB file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

createSMBFileShare_authentication :: Lens' CreateSMBFileShare (Maybe Text)

The authentication method that users use to access the file share. The default is ActiveDirectory.

Valid Values: ActiveDirectory | GuestAccess

createSMBFileShare_cacheAttributes :: Lens' CreateSMBFileShare (Maybe CacheAttributes)

Specifies refresh cache information for the file share.

createSMBFileShare_objectACL :: Lens' CreateSMBFileShare (Maybe ObjectACL)

A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

createSMBFileShare_kmsEncrypted :: Lens' CreateSMBFileShare (Maybe Bool)

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

createSMBFileShare_defaultStorageClass :: Lens' CreateSMBFileShare (Maybe Text)

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

createSMBFileShare_fileShareName :: Lens' CreateSMBFileShare (Maybe Text)

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN, or if an access point or access point alias is used.

createSMBFileShare_sMBACLEnabled :: Lens' CreateSMBFileShare (Maybe Bool)

Set this value to true to enable access control list (ACL) on the SMB file share. Set it to false to map file and directory permissions to the POSIX permissions.

For more information, see Using Microsoft Windows ACLs to control access to an SMB file share in the Storage Gateway User Guide.

Valid Values: true | false

createSMBFileShare_oplocksEnabled :: Lens' CreateSMBFileShare (Maybe Bool)

Specifies whether opportunistic locking is enabled for the SMB file share.

Enabling opportunistic locking on case-sensitive shares is not recommended for workloads that involve access to files with the same name in different case.

Valid Values: true | false

createSMBFileShare_notificationPolicy :: Lens' CreateSMBFileShare (Maybe Text)

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

createSMBFileShare_requesterPays :: Lens' CreateSMBFileShare (Maybe Bool)

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

createSMBFileShare_guessMIMETypeEnabled :: Lens' CreateSMBFileShare (Maybe Bool)

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

createSMBFileShare_readOnly :: Lens' CreateSMBFileShare (Maybe Bool)

A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

createSMBFileShare_bucketRegion :: Lens' CreateSMBFileShare (Maybe Text)

Specifies the Region of the S3 bucket where the SMB file share stores files.

This parameter is required for SMB file shares that connect to Amazon S3 through a VPC endpoint, a VPC access point, or an access point alias that points to a VPC access point.

createSMBFileShare_caseSensitivity :: Lens' CreateSMBFileShare (Maybe CaseSensitivity)

The case of an object name in an Amazon S3 bucket. For ClientSpecified, the client determines the case sensitivity. For CaseSensitive, the gateway determines the case sensitivity. The default value is ClientSpecified.

createSMBFileShare_tags :: Lens' CreateSMBFileShare (Maybe [Tag])

A list of up to 50 tags that can be assigned to the NFS file share. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

createSMBFileShare_clientToken :: Lens' CreateSMBFileShare Text

A unique string value that you supply that is used by S3 File Gateway to ensure idempotent file share creation.

createSMBFileShare_gatewayARN :: Lens' CreateSMBFileShare Text

The ARN of the S3 File Gateway on which you want to create a file share.

createSMBFileShare_role :: Lens' CreateSMBFileShare Text

The ARN of the Identity and Access Management (IAM) role that an S3 File Gateway assumes when it accesses the underlying storage.

createSMBFileShare_locationARN :: Lens' CreateSMBFileShare Text

A custom ARN for the backend storage used for storing data for file shares. It includes a resource ARN with an optional prefix concatenation. The prefix must end with a forward slash (/).

You can specify LocationARN as a bucket ARN, access point ARN or access point alias, as shown in the following examples.

Bucket ARN:

arn:aws:s3:::my-bucket/prefix/

Access point ARN:

arn:aws:s3:region:account-id:accesspoint/access-point-name/prefix/

If you specify an access point, the bucket policy must be configured to delegate access control to the access point. For information, see Delegating access control to access points in the Amazon S3 User Guide.

Access point alias:

test-ap-ab123cdef4gehijklmn5opqrstuvuse1a-s3alias

Destructuring the Response

data CreateSMBFileShareResponse

CreateSMBFileShareOutput

See: newCreateSMBFileShareResponse smart constructor.

Constructors

CreateSMBFileShareResponse'

Fields

fileShareARN :: Maybe Text

The Amazon Resource Name (ARN) of the newly created file share.

httpStatus :: Int

The response's http status code.

newCreateSMBFileShareResponse

Arguments

:: Int
$sel:httpStatus:CreateSMBFileShareResponse'

-> CreateSMBFileShareResponse

Create a value of CreateSMBFileShareResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileShareARN:CreateSMBFileShareResponse', createSMBFileShareResponse_fileShareARN - The Amazon Resource Name (ARN) of the newly created file share.

$sel:httpStatus:CreateSMBFileShareResponse', createSMBFileShareResponse_httpStatus - The response's http status code.

Response Lenses

createSMBFileShareResponse_fileShareARN :: Lens' CreateSMBFileShareResponse (Maybe Text)

The Amazon Resource Name (ARN) of the newly created file share.

createSMBFileShareResponse_httpStatus :: Lens' CreateSMBFileShareResponse Int

The response's http status code.