Amazonka.StorageGateway.DescribeCache

Description

Returns information about the cache of a gateway. This operation is only supported in the cached volume, tape, and file gateway types.

The response includes disk IDs that are configured as cache, and it includes the amount of cache allocated and used.

Creating a Request

data DescribeCache

See: newDescribeCache smart constructor.

Constructors

DescribeCache'

Fields

gatewayARN :: Text

newDescribeCache

Arguments

:: Text
$sel:gatewayARN:DescribeCache'

-> DescribeCache

Create a value of DescribeCache with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:DescribeCache', describeCache_gatewayARN - Undocumented member.

Request Lenses

describeCache_gatewayARN :: Lens' DescribeCache Text

Undocumented member.

Destructuring the Response

data DescribeCacheResponse

See: newDescribeCacheResponse smart constructor.

Constructors

DescribeCacheResponse'

Fields

gatewayARN :: Maybe Text

diskIds :: Maybe [Text]

An array of strings that identify disks that are to be configured as working storage. Each string has a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

cacheUsedPercentage :: Maybe Double

Percent use of the gateway's cache storage. This metric applies only to the gateway-cached volume setup. The sample is taken at the end of the reporting period.

cacheHitPercentage :: Maybe Double

Percent of application read operations from the file shares that are served from cache. The sample is taken at the end of the reporting period.

cacheMissPercentage :: Maybe Double

Percent of application read operations from the file shares that are not served from cache. The sample is taken at the end of the reporting period.

cacheAllocatedInBytes :: Maybe Integer

The amount of cache in bytes allocated to a gateway.

cacheDirtyPercentage :: Maybe Double

The file share's contribution to the overall percentage of the gateway's cache that has not been persisted to Amazon Web Services. The sample is taken at the end of the reporting period.

httpStatus :: Int

The response's http status code.

newDescribeCacheResponse

Arguments

:: Int
$sel:httpStatus:DescribeCacheResponse'

-> DescribeCacheResponse

Create a value of DescribeCacheResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:DescribeCache', describeCacheResponse_gatewayARN - Undocumented member.

$sel:diskIds:DescribeCacheResponse', describeCacheResponse_diskIds - An array of strings that identify disks that are to be configured as working storage. Each string has a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

$sel:cacheUsedPercentage:DescribeCacheResponse', describeCacheResponse_cacheUsedPercentage - Percent use of the gateway's cache storage. This metric applies only to the gateway-cached volume setup. The sample is taken at the end of the reporting period.

$sel:cacheHitPercentage:DescribeCacheResponse', describeCacheResponse_cacheHitPercentage - Percent of application read operations from the file shares that are served from cache. The sample is taken at the end of the reporting period.

$sel:cacheMissPercentage:DescribeCacheResponse', describeCacheResponse_cacheMissPercentage - Percent of application read operations from the file shares that are not served from cache. The sample is taken at the end of the reporting period.

$sel:cacheAllocatedInBytes:DescribeCacheResponse', describeCacheResponse_cacheAllocatedInBytes - The amount of cache in bytes allocated to a gateway.

$sel:cacheDirtyPercentage:DescribeCacheResponse', describeCacheResponse_cacheDirtyPercentage - The file share's contribution to the overall percentage of the gateway's cache that has not been persisted to Amazon Web Services. The sample is taken at the end of the reporting period.

$sel:httpStatus:DescribeCacheResponse', describeCacheResponse_httpStatus - The response's http status code.

Response Lenses

describeCacheResponse_gatewayARN :: Lens' DescribeCacheResponse (Maybe Text)

Undocumented member.

describeCacheResponse_diskIds :: Lens' DescribeCacheResponse (Maybe [Text])

An array of strings that identify disks that are to be configured as working storage. Each string has a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

describeCacheResponse_cacheUsedPercentage :: Lens' DescribeCacheResponse (Maybe Double)

Percent use of the gateway's cache storage. This metric applies only to the gateway-cached volume setup. The sample is taken at the end of the reporting period.

describeCacheResponse_cacheHitPercentage :: Lens' DescribeCacheResponse (Maybe Double)

Percent of application read operations from the file shares that are served from cache. The sample is taken at the end of the reporting period.

describeCacheResponse_cacheMissPercentage :: Lens' DescribeCacheResponse (Maybe Double)

Percent of application read operations from the file shares that are not served from cache. The sample is taken at the end of the reporting period.

describeCacheResponse_cacheAllocatedInBytes :: Lens' DescribeCacheResponse (Maybe Integer)

The amount of cache in bytes allocated to a gateway.

describeCacheResponse_cacheDirtyPercentage :: Lens' DescribeCacheResponse (Maybe Double)

The file share's contribution to the overall percentage of the gateway's cache that has not been persisted to Amazon Web Services. The sample is taken at the end of the reporting period.

describeCacheResponse_httpStatus :: Lens' DescribeCacheResponse Int

The response's http status code.