Amazonka.StorageGateway.DescribeTapeArchives

Description

Returns a description of specified virtual tapes in the virtual tape shelf (VTS). This operation is only supported in the tape gateway type.

If a specific TapeARN is not specified, Storage Gateway returns a description of all virtual tapes found in the VTS associated with your account.

This operation returns paginated results.

Creating a Request

data DescribeTapeArchives

DescribeTapeArchivesInput

See: newDescribeTapeArchives smart constructor.

Constructors

DescribeTapeArchives'

Fields

marker :: Maybe Text

An opaque string that indicates the position at which to begin describing virtual tapes.

limit :: Maybe Natural

Specifies that the number of virtual tapes described be limited to the specified number.

tapeARNs :: Maybe [Text]

Specifies one or more unique Amazon Resource Names (ARNs) that represent the virtual tapes you want to describe.

newDescribeTapeArchives :: DescribeTapeArchives

Create a value of DescribeTapeArchives with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeTapeArchives', describeTapeArchives_marker - An opaque string that indicates the position at which to begin describing virtual tapes.

$sel:limit:DescribeTapeArchives', describeTapeArchives_limit - Specifies that the number of virtual tapes described be limited to the specified number.

$sel:tapeARNs:DescribeTapeArchives', describeTapeArchives_tapeARNs - Specifies one or more unique Amazon Resource Names (ARNs) that represent the virtual tapes you want to describe.

Request Lenses

describeTapeArchives_marker :: Lens' DescribeTapeArchives (Maybe Text)

An opaque string that indicates the position at which to begin describing virtual tapes.

describeTapeArchives_limit :: Lens' DescribeTapeArchives (Maybe Natural)

Specifies that the number of virtual tapes described be limited to the specified number.

describeTapeArchives_tapeARNs :: Lens' DescribeTapeArchives (Maybe [Text])

Specifies one or more unique Amazon Resource Names (ARNs) that represent the virtual tapes you want to describe.

Destructuring the Response

data DescribeTapeArchivesResponse

DescribeTapeArchivesOutput

See: newDescribeTapeArchivesResponse smart constructor.

Constructors

DescribeTapeArchivesResponse'

Fields

tapeArchives :: Maybe [TapeArchive]

An array of virtual tape objects in the virtual tape shelf (VTS). The description includes of the Amazon Resource Name (ARN) of the virtual tapes. The information returned includes the Amazon Resource Names (ARNs) of the tapes, size of the tapes, status of the tapes, progress of the description, and tape barcode.

marker :: Maybe Text

An opaque string that indicates the position at which the virtual tapes that were fetched for description ended. Use this marker in your next request to fetch the next set of virtual tapes in the virtual tape shelf (VTS). If there are no more virtual tapes to describe, this field does not appear in the response.

httpStatus :: Int

The response's http status code.

newDescribeTapeArchivesResponse

Arguments

:: Int
$sel:httpStatus:DescribeTapeArchivesResponse'

-> DescribeTapeArchivesResponse

Create a value of DescribeTapeArchivesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tapeArchives:DescribeTapeArchivesResponse', describeTapeArchivesResponse_tapeArchives - An array of virtual tape objects in the virtual tape shelf (VTS). The description includes of the Amazon Resource Name (ARN) of the virtual tapes. The information returned includes the Amazon Resource Names (ARNs) of the tapes, size of the tapes, status of the tapes, progress of the description, and tape barcode.

$sel:marker:DescribeTapeArchives', describeTapeArchivesResponse_marker - An opaque string that indicates the position at which the virtual tapes that were fetched for description ended. Use this marker in your next request to fetch the next set of virtual tapes in the virtual tape shelf (VTS). If there are no more virtual tapes to describe, this field does not appear in the response.

$sel:httpStatus:DescribeTapeArchivesResponse', describeTapeArchivesResponse_httpStatus - The response's http status code.

Response Lenses

describeTapeArchivesResponse_tapeArchives :: Lens' DescribeTapeArchivesResponse (Maybe [TapeArchive])

An array of virtual tape objects in the virtual tape shelf (VTS). The description includes of the Amazon Resource Name (ARN) of the virtual tapes. The information returned includes the Amazon Resource Names (ARNs) of the tapes, size of the tapes, status of the tapes, progress of the description, and tape barcode.

describeTapeArchivesResponse_marker :: Lens' DescribeTapeArchivesResponse (Maybe Text)

An opaque string that indicates the position at which the virtual tapes that were fetched for description ended. Use this marker in your next request to fetch the next set of virtual tapes in the virtual tape shelf (VTS). If there are no more virtual tapes to describe, this field does not appear in the response.

describeTapeArchivesResponse_httpStatus :: Lens' DescribeTapeArchivesResponse Int

The response's http status code.