Amazonka.StorageGateway.DescribeWorkingStorage

Description

Returns information about the working storage of a gateway. This operation is only supported in the stored volumes gateway type. This operation is deprecated in cached volumes API version (20120630). Use DescribeUploadBuffer instead.

Working storage is also referred to as upload buffer. You can also use the DescribeUploadBuffer operation to add upload buffer to a stored volume gateway.

The response includes disk IDs that are configured as working storage, and it includes the amount of working storage allocated and used.

Creating a Request

data DescribeWorkingStorage

A JSON object containing the Amazon Resource Name (ARN) of the gateway.

See: newDescribeWorkingStorage smart constructor.

Constructors

DescribeWorkingStorage'

Fields

gatewayARN :: Text

newDescribeWorkingStorage

Arguments

:: Text
$sel:gatewayARN:DescribeWorkingStorage'

-> DescribeWorkingStorage

Create a value of DescribeWorkingStorage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:DescribeWorkingStorage', describeWorkingStorage_gatewayARN - Undocumented member.

Request Lenses

describeWorkingStorage_gatewayARN :: Lens' DescribeWorkingStorage Text

Undocumented member.

Destructuring the Response

data DescribeWorkingStorageResponse

A JSON object containing the following fields:

See: newDescribeWorkingStorageResponse smart constructor.

Constructors

DescribeWorkingStorageResponse'

Fields

gatewayARN :: Maybe Text

diskIds :: Maybe [Text]

An array of the gateway's local disk IDs that are configured as working storage. Each local disk ID is specified as a string (minimum length of 1 and maximum length of 300). If no local disks are configured as working storage, then the DiskIds array is empty.

workingStorageAllocatedInBytes :: Maybe Integer

The total working storage in bytes allocated for the gateway. If no working storage is configured for the gateway, this field returns 0.

workingStorageUsedInBytes :: Maybe Integer

The total working storage in bytes in use by the gateway. If no working storage is configured for the gateway, this field returns 0.

httpStatus :: Int

The response's http status code.

newDescribeWorkingStorageResponse

Arguments

:: Int
$sel:httpStatus:DescribeWorkingStorageResponse'

-> DescribeWorkingStorageResponse

Create a value of DescribeWorkingStorageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:DescribeWorkingStorage', describeWorkingStorageResponse_gatewayARN - Undocumented member.

$sel:diskIds:DescribeWorkingStorageResponse', describeWorkingStorageResponse_diskIds - An array of the gateway's local disk IDs that are configured as working storage. Each local disk ID is specified as a string (minimum length of 1 and maximum length of 300). If no local disks are configured as working storage, then the DiskIds array is empty.

$sel:workingStorageAllocatedInBytes:DescribeWorkingStorageResponse', describeWorkingStorageResponse_workingStorageAllocatedInBytes - The total working storage in bytes allocated for the gateway. If no working storage is configured for the gateway, this field returns 0.

$sel:workingStorageUsedInBytes:DescribeWorkingStorageResponse', describeWorkingStorageResponse_workingStorageUsedInBytes - The total working storage in bytes in use by the gateway. If no working storage is configured for the gateway, this field returns 0.

$sel:httpStatus:DescribeWorkingStorageResponse', describeWorkingStorageResponse_httpStatus - The response's http status code.

Response Lenses

describeWorkingStorageResponse_gatewayARN :: Lens' DescribeWorkingStorageResponse (Maybe Text)

Undocumented member.

describeWorkingStorageResponse_diskIds :: Lens' DescribeWorkingStorageResponse (Maybe [Text])

An array of the gateway's local disk IDs that are configured as working storage. Each local disk ID is specified as a string (minimum length of 1 and maximum length of 300). If no local disks are configured as working storage, then the DiskIds array is empty.

describeWorkingStorageResponse_workingStorageAllocatedInBytes :: Lens' DescribeWorkingStorageResponse (Maybe Integer)

The total working storage in bytes allocated for the gateway. If no working storage is configured for the gateway, this field returns 0.

describeWorkingStorageResponse_workingStorageUsedInBytes :: Lens' DescribeWorkingStorageResponse (Maybe Integer)

The total working storage in bytes in use by the gateway. If no working storage is configured for the gateway, this field returns 0.

describeWorkingStorageResponse_httpStatus :: Lens' DescribeWorkingStorageResponse Int

The response's http status code.