Amazonka.StorageGateway.ListGateways

Description

Lists gateways owned by an Amazon Web Services account in an Amazon Web Services Region specified in the request. The returned list is ordered by gateway Amazon Resource Name (ARN).

By default, the operation returns a maximum of 100 gateways. This operation supports pagination that allows you to optionally reduce the number of gateways returned in a response.

If you have more gateways than are returned in a response (that is, the response returns only a truncated list of your gateways), the response contains a marker that you can specify in your next request to fetch the next page of gateways.

This operation returns paginated results.

Creating a Request

data ListGateways

A JSON object containing zero or more of the following fields:

ListGatewaysInput$Limit

ListGatewaysInput$Marker

See: newListGateways smart constructor.

Constructors

ListGateways'

Fields

marker :: Maybe Text

An opaque string that indicates the position at which to begin the returned list of gateways.

limit :: Maybe Natural

Specifies that the list of gateways returned be limited to the specified number of items.

newListGateways :: ListGateways

Create a value of ListGateways with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListGateways', listGateways_marker - An opaque string that indicates the position at which to begin the returned list of gateways.

$sel:limit:ListGateways', listGateways_limit - Specifies that the list of gateways returned be limited to the specified number of items.

Request Lenses

listGateways_marker :: Lens' ListGateways (Maybe Text)

An opaque string that indicates the position at which to begin the returned list of gateways.

listGateways_limit :: Lens' ListGateways (Maybe Natural)

Specifies that the list of gateways returned be limited to the specified number of items.

Destructuring the Response

data ListGatewaysResponse

See: newListGatewaysResponse smart constructor.

Constructors

ListGatewaysResponse'

Fields

marker :: Maybe Text

Use the marker in your next request to fetch the next set of gateways in the list. If there are no more gateways to list, this field does not appear in the response.

gateways :: Maybe [GatewayInfo]

An array of GatewayInfo objects.

httpStatus :: Int

The response's http status code.

newListGatewaysResponse

Arguments

:: Int
$sel:httpStatus:ListGatewaysResponse'

-> ListGatewaysResponse

Create a value of ListGatewaysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListGateways', listGatewaysResponse_marker - Use the marker in your next request to fetch the next set of gateways in the list. If there are no more gateways to list, this field does not appear in the response.

$sel:gateways:ListGatewaysResponse', listGatewaysResponse_gateways - An array of GatewayInfo objects.

$sel:httpStatus:ListGatewaysResponse', listGatewaysResponse_httpStatus - The response's http status code.

Response Lenses

listGatewaysResponse_marker :: Lens' ListGatewaysResponse (Maybe Text)

Use the marker in your next request to fetch the next set of gateways in the list. If there are no more gateways to list, this field does not appear in the response.

listGatewaysResponse_gateways :: Lens' ListGatewaysResponse (Maybe [GatewayInfo])

An array of GatewayInfo objects.

listGatewaysResponse_httpStatus :: Lens' ListGatewaysResponse Int

The response's http status code.