Amazonka.StorageGateway.ListTapePools

Description

Lists custom tape pools. You specify custom tape pools to list by specifying one or more custom tape pool Amazon Resource Names (ARNs). If you don't specify a custom tape pool ARN, the operation lists all custom tape pools.

This operation supports pagination. You can optionally specify the Limit parameter in the body to limit the number of tape pools in the response. If the number of tape pools returned in the response is truncated, the response includes a Marker element that you can use in your subsequent request to retrieve the next set of tape pools.

This operation returns paginated results.

Creating a Request

data ListTapePools

See: newListTapePools smart constructor.

Constructors

ListTapePools'

Fields

poolARNs :: Maybe [Text]

The Amazon Resource Name (ARN) of each of the custom tape pools you want to list. If you don't specify a custom tape pool ARN, the response lists all custom tape pools.

marker :: Maybe Text

A string that indicates the position at which to begin the returned list of tape pools.

limit :: Maybe Natural

An optional number limit for the tape pools in the list returned by this call.

newListTapePools :: ListTapePools

Create a value of ListTapePools with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:poolARNs:ListTapePools', listTapePools_poolARNs - The Amazon Resource Name (ARN) of each of the custom tape pools you want to list. If you don't specify a custom tape pool ARN, the response lists all custom tape pools.

$sel:marker:ListTapePools', listTapePools_marker - A string that indicates the position at which to begin the returned list of tape pools.

$sel:limit:ListTapePools', listTapePools_limit - An optional number limit for the tape pools in the list returned by this call.

Request Lenses

listTapePools_poolARNs :: Lens' ListTapePools (Maybe [Text])

The Amazon Resource Name (ARN) of each of the custom tape pools you want to list. If you don't specify a custom tape pool ARN, the response lists all custom tape pools.

listTapePools_marker :: Lens' ListTapePools (Maybe Text)

A string that indicates the position at which to begin the returned list of tape pools.

listTapePools_limit :: Lens' ListTapePools (Maybe Natural)

An optional number limit for the tape pools in the list returned by this call.

Destructuring the Response

data ListTapePoolsResponse

See: newListTapePoolsResponse smart constructor.

Constructors

ListTapePoolsResponse'

Fields

poolInfos :: Maybe [PoolInfo]

An array of PoolInfo objects, where each object describes a single custom tape pool. If there are no custom tape pools, the PoolInfos is an empty array.

marker :: Maybe Text

A string that indicates the position at which to begin the returned list of tape pools. Use the marker in your next request to continue pagination of tape pools. If there are no more tape pools to list, this element does not appear in the response body.

httpStatus :: Int

The response's http status code.

newListTapePoolsResponse

Arguments

:: Int
$sel:httpStatus:ListTapePoolsResponse'

-> ListTapePoolsResponse

Create a value of ListTapePoolsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:poolInfos:ListTapePoolsResponse', listTapePoolsResponse_poolInfos - An array of PoolInfo objects, where each object describes a single custom tape pool. If there are no custom tape pools, the PoolInfos is an empty array.

$sel:marker:ListTapePools', listTapePoolsResponse_marker - A string that indicates the position at which to begin the returned list of tape pools. Use the marker in your next request to continue pagination of tape pools. If there are no more tape pools to list, this element does not appear in the response body.

$sel:httpStatus:ListTapePoolsResponse', listTapePoolsResponse_httpStatus - The response's http status code.

Response Lenses

listTapePoolsResponse_poolInfos :: Lens' ListTapePoolsResponse (Maybe [PoolInfo])

An array of PoolInfo objects, where each object describes a single custom tape pool. If there are no custom tape pools, the PoolInfos is an empty array.

listTapePoolsResponse_marker :: Lens' ListTapePoolsResponse (Maybe Text)

A string that indicates the position at which to begin the returned list of tape pools. Use the marker in your next request to continue pagination of tape pools. If there are no more tape pools to list, this element does not appear in the response body.

listTapePoolsResponse_httpStatus :: Lens' ListTapePoolsResponse Int

The response's http status code.