Amazonka.StorageGateway.ListTapes

Description

Lists virtual tapes in your virtual tape library (VTL) and your virtual tape shelf (VTS). You specify the tapes to list by specifying one or more tape Amazon Resource Names (ARNs). If you don't specify a tape ARN, the operation lists all virtual tapes in both your VTL and VTS.

This operation supports pagination. By default, the operation returns a maximum of up to 100 tapes. You can optionally specify the Limit parameter in the body to limit the number of tapes in the response. If the number of tapes returned in the response is truncated, the response includes a Marker element that you can use in your subsequent request to retrieve the next set of tapes. This operation is only supported in the tape gateway type.

This operation returns paginated results.

Creating a Request

data ListTapes

A JSON object that contains one or more of the following fields:

ListTapesInput$Limit

ListTapesInput$Marker

ListTapesInput$TapeARNs

See: newListTapes smart constructor.

Constructors

ListTapes'

Fields

marker :: Maybe Text

A string that indicates the position at which to begin the returned list of tapes.

limit :: Maybe Natural

An optional number limit for the tapes in the list returned by this call.

tapeARNs :: Maybe [Text]

newListTapes :: ListTapes

Create a value of ListTapes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListTapes', listTapes_marker - A string that indicates the position at which to begin the returned list of tapes.

$sel:limit:ListTapes', listTapes_limit - An optional number limit for the tapes in the list returned by this call.

$sel:tapeARNs:ListTapes', listTapes_tapeARNs - Undocumented member.

Request Lenses

listTapes_marker :: Lens' ListTapes (Maybe Text)

A string that indicates the position at which to begin the returned list of tapes.

listTapes_limit :: Lens' ListTapes (Maybe Natural)

An optional number limit for the tapes in the list returned by this call.

listTapes_tapeARNs :: Lens' ListTapes (Maybe [Text])

Undocumented member.

Destructuring the Response

data ListTapesResponse

A JSON object containing the following fields:

ListTapesOutput$Marker

ListTapesOutput$VolumeInfos

See: newListTapesResponse smart constructor.

Constructors

ListTapesResponse'

Fields

marker :: Maybe Text

A string that indicates the position at which to begin returning the next list of tapes. Use the marker in your next request to continue pagination of tapes. If there are no more tapes to list, this element does not appear in the response body.

tapeInfos :: Maybe [TapeInfo]

httpStatus :: Int

The response's http status code.

newListTapesResponse

Arguments

:: Int
$sel:httpStatus:ListTapesResponse'

-> ListTapesResponse

Create a value of ListTapesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListTapes', listTapesResponse_marker - A string that indicates the position at which to begin returning the next list of tapes. Use the marker in your next request to continue pagination of tapes. If there are no more tapes to list, this element does not appear in the response body.

$sel:tapeInfos:ListTapesResponse', listTapesResponse_tapeInfos - Undocumented member.

$sel:httpStatus:ListTapesResponse', listTapesResponse_httpStatus - The response's http status code.

Response Lenses

listTapesResponse_marker :: Lens' ListTapesResponse (Maybe Text)

A string that indicates the position at which to begin returning the next list of tapes. Use the marker in your next request to continue pagination of tapes. If there are no more tapes to list, this element does not appear in the response body.

listTapesResponse_tapeInfos :: Lens' ListTapesResponse (Maybe [TapeInfo])

Undocumented member.

listTapesResponse_httpStatus :: Lens' ListTapesResponse Int

The response's http status code.