Amazonka.StorageGateway.Types.StorediSCSIVolume

Description

Documentation

data StorediSCSIVolume

Describes an iSCSI stored volume.

See: newStorediSCSIVolume smart constructor.

Constructors

StorediSCSIVolume'

Fields

volumeiSCSIAttributes :: Maybe VolumeiSCSIAttributes

An VolumeiSCSIAttributes object that represents a collection of iSCSI attributes for one stored volume.

volumeStatus :: Maybe Text

One of the VolumeStatus values that indicates the state of the storage volume.

sourceSnapshotId :: Maybe Text

If the stored volume was created from a snapshot, this field contains the snapshot ID used, e.g. snap-78e22663. Otherwise, this field is not included.

preservedExistingData :: Maybe Bool

Indicates if when the stored volume was created, existing data on the underlying local disk was preserved.

Valid Values: true | false

kmsKey :: Maybe Text

volumeAttachmentStatus :: Maybe Text

A value that indicates whether a storage volume is attached to, detached from, or is in the process of detaching from a gateway. For more information, see Moving your volumes to a different gateway.

volumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the storage volume.

volumeProgress :: Maybe Double

Represents the percentage complete if the volume is restoring or bootstrapping that represents the percent of data transferred. This field does not appear in the response if the stored volume is not restoring or bootstrapping.

volumeSizeInBytes :: Maybe Integer

The size of the volume in bytes.

volumeUsedInBytes :: Maybe Integer

The size of the data stored on the volume in bytes. This value is calculated based on the number of blocks that are touched, instead of the actual amount of data written. This value can be useful for sequential write patterns but less accurate for random write patterns. VolumeUsedInBytes is different from the compressed size of the volume, which is the value that is used to calculate your bill.

This value is not available for volumes created prior to May 13, 2015, until you store data on the volume.

createdDate :: Maybe POSIX

The date the volume was created. Volumes created prior to March 28, 2017 don’t have this timestamp.

volumeId :: Maybe Text

The unique identifier of the volume, e.g., vol-AE4B946D.

volumeDiskId :: Maybe Text

The ID of the local disk that was specified in the CreateStorediSCSIVolume operation.

volumeType :: Maybe Text

One of the VolumeType enumeration values describing the type of the volume.

targetName :: Maybe Text

The name of the iSCSI target used by an initiator to connect to a volume and used as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes on a gateway.

If you don't specify a value, Storage Gateway uses the value that was previously used for this volume as the new target name.

newStorediSCSIVolume :: StorediSCSIVolume

Create a value of StorediSCSIVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeiSCSIAttributes:StorediSCSIVolume', storediSCSIVolume_volumeiSCSIAttributes - An VolumeiSCSIAttributes object that represents a collection of iSCSI attributes for one stored volume.

$sel:volumeStatus:StorediSCSIVolume', storediSCSIVolume_volumeStatus - One of the VolumeStatus values that indicates the state of the storage volume.

$sel:sourceSnapshotId:StorediSCSIVolume', storediSCSIVolume_sourceSnapshotId - If the stored volume was created from a snapshot, this field contains the snapshot ID used, e.g. snap-78e22663. Otherwise, this field is not included.

$sel:preservedExistingData:StorediSCSIVolume', storediSCSIVolume_preservedExistingData - Indicates if when the stored volume was created, existing data on the underlying local disk was preserved.

Valid Values: true | false

$sel:kmsKey:StorediSCSIVolume', storediSCSIVolume_kmsKey - Undocumented member.

$sel:volumeAttachmentStatus:StorediSCSIVolume', storediSCSIVolume_volumeAttachmentStatus - A value that indicates whether a storage volume is attached to, detached from, or is in the process of detaching from a gateway. For more information, see Moving your volumes to a different gateway.

$sel:volumeARN:StorediSCSIVolume', storediSCSIVolume_volumeARN - The Amazon Resource Name (ARN) of the storage volume.

$sel:volumeProgress:StorediSCSIVolume', storediSCSIVolume_volumeProgress - Represents the percentage complete if the volume is restoring or bootstrapping that represents the percent of data transferred. This field does not appear in the response if the stored volume is not restoring or bootstrapping.

$sel:volumeSizeInBytes:StorediSCSIVolume', storediSCSIVolume_volumeSizeInBytes - The size of the volume in bytes.

$sel:volumeUsedInBytes:StorediSCSIVolume', storediSCSIVolume_volumeUsedInBytes - The size of the data stored on the volume in bytes. This value is calculated based on the number of blocks that are touched, instead of the actual amount of data written. This value can be useful for sequential write patterns but less accurate for random write patterns. VolumeUsedInBytes is different from the compressed size of the volume, which is the value that is used to calculate your bill.

This value is not available for volumes created prior to May 13, 2015, until you store data on the volume.

$sel:createdDate:StorediSCSIVolume', storediSCSIVolume_createdDate - The date the volume was created. Volumes created prior to March 28, 2017 don’t have this timestamp.

$sel:volumeId:StorediSCSIVolume', storediSCSIVolume_volumeId - The unique identifier of the volume, e.g., vol-AE4B946D.

$sel:volumeDiskId:StorediSCSIVolume', storediSCSIVolume_volumeDiskId - The ID of the local disk that was specified in the CreateStorediSCSIVolume operation.

$sel:volumeType:StorediSCSIVolume', storediSCSIVolume_volumeType - One of the VolumeType enumeration values describing the type of the volume.

$sel:targetName:StorediSCSIVolume', storediSCSIVolume_targetName - The name of the iSCSI target used by an initiator to connect to a volume and used as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes on a gateway.

If you don't specify a value, Storage Gateway uses the value that was previously used for this volume as the new target name.

storediSCSIVolume_volumeiSCSIAttributes :: Lens' StorediSCSIVolume (Maybe VolumeiSCSIAttributes)

An VolumeiSCSIAttributes object that represents a collection of iSCSI attributes for one stored volume.

storediSCSIVolume_volumeStatus :: Lens' StorediSCSIVolume (Maybe Text)

One of the VolumeStatus values that indicates the state of the storage volume.

storediSCSIVolume_sourceSnapshotId :: Lens' StorediSCSIVolume (Maybe Text)

If the stored volume was created from a snapshot, this field contains the snapshot ID used, e.g. snap-78e22663. Otherwise, this field is not included.

storediSCSIVolume_preservedExistingData :: Lens' StorediSCSIVolume (Maybe Bool)

Indicates if when the stored volume was created, existing data on the underlying local disk was preserved.

Valid Values: true | false

storediSCSIVolume_kmsKey :: Lens' StorediSCSIVolume (Maybe Text)

Undocumented member.

storediSCSIVolume_volumeAttachmentStatus :: Lens' StorediSCSIVolume (Maybe Text)

A value that indicates whether a storage volume is attached to, detached from, or is in the process of detaching from a gateway. For more information, see Moving your volumes to a different gateway.

storediSCSIVolume_volumeARN :: Lens' StorediSCSIVolume (Maybe Text)

The Amazon Resource Name (ARN) of the storage volume.

storediSCSIVolume_volumeProgress :: Lens' StorediSCSIVolume (Maybe Double)

Represents the percentage complete if the volume is restoring or bootstrapping that represents the percent of data transferred. This field does not appear in the response if the stored volume is not restoring or bootstrapping.

storediSCSIVolume_volumeSizeInBytes :: Lens' StorediSCSIVolume (Maybe Integer)

The size of the volume in bytes.

storediSCSIVolume_volumeUsedInBytes :: Lens' StorediSCSIVolume (Maybe Integer)

The size of the data stored on the volume in bytes. This value is calculated based on the number of blocks that are touched, instead of the actual amount of data written. This value can be useful for sequential write patterns but less accurate for random write patterns. VolumeUsedInBytes is different from the compressed size of the volume, which is the value that is used to calculate your bill.

This value is not available for volumes created prior to May 13, 2015, until you store data on the volume.

storediSCSIVolume_createdDate :: Lens' StorediSCSIVolume (Maybe UTCTime)

The date the volume was created. Volumes created prior to March 28, 2017 don’t have this timestamp.

storediSCSIVolume_volumeId :: Lens' StorediSCSIVolume (Maybe Text)

The unique identifier of the volume, e.g., vol-AE4B946D.

storediSCSIVolume_volumeDiskId :: Lens' StorediSCSIVolume (Maybe Text)

The ID of the local disk that was specified in the CreateStorediSCSIVolume operation.

storediSCSIVolume_volumeType :: Lens' StorediSCSIVolume (Maybe Text)

One of the VolumeType enumeration values describing the type of the volume.

storediSCSIVolume_targetName :: Lens' StorediSCSIVolume (Maybe Text)

The name of the iSCSI target used by an initiator to connect to a volume and used as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes on a gateway.

If you don't specify a value, Storage Gateway uses the value that was previously used for this volume as the new target name.