Amazonka.StorageGateway.Types.TapeRecoveryPointInfo

Description

Documentation

data TapeRecoveryPointInfo

Describes a recovery point.

See: newTapeRecoveryPointInfo smart constructor.

Constructors

TapeRecoveryPointInfo'

Fields

tapeStatus :: Maybe Text

The status of the virtual tapes.

tapeRecoveryPointTime :: Maybe POSIX

The time when the point-in-time view of the virtual tape was replicated for later recovery.

The default timestamp format of the tape recovery point time is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

tapeARN :: Maybe Text

The Amazon Resource Name (ARN) of the virtual tape.

tapeSizeInBytes :: Maybe Integer

The size, in bytes, of the virtual tapes to recover.

newTapeRecoveryPointInfo :: TapeRecoveryPointInfo

Create a value of TapeRecoveryPointInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tapeStatus:TapeRecoveryPointInfo', tapeRecoveryPointInfo_tapeStatus - The status of the virtual tapes.

$sel:tapeRecoveryPointTime:TapeRecoveryPointInfo', tapeRecoveryPointInfo_tapeRecoveryPointTime - The time when the point-in-time view of the virtual tape was replicated for later recovery.

The default timestamp format of the tape recovery point time is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

$sel:tapeARN:TapeRecoveryPointInfo', tapeRecoveryPointInfo_tapeARN - The Amazon Resource Name (ARN) of the virtual tape.

$sel:tapeSizeInBytes:TapeRecoveryPointInfo', tapeRecoveryPointInfo_tapeSizeInBytes - The size, in bytes, of the virtual tapes to recover.

tapeRecoveryPointInfo_tapeStatus :: Lens' TapeRecoveryPointInfo (Maybe Text)

The status of the virtual tapes.

tapeRecoveryPointInfo_tapeRecoveryPointTime :: Lens' TapeRecoveryPointInfo (Maybe UTCTime)

The time when the point-in-time view of the virtual tape was replicated for later recovery.

The default timestamp format of the tape recovery point time is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

tapeRecoveryPointInfo_tapeARN :: Lens' TapeRecoveryPointInfo (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape.

tapeRecoveryPointInfo_tapeSizeInBytes :: Lens' TapeRecoveryPointInfo (Maybe Integer)

The size, in bytes, of the virtual tapes to recover.