Amazonka.StorageGateway.Types.VolumeInfo

Description

Documentation

data VolumeInfo

Describes a storage volume object.

See: newVolumeInfo smart constructor.

Constructors

VolumeInfo'

Fields

gatewayARN :: Maybe Text

volumeAttachmentStatus :: Maybe Text

One of the VolumeStatus values that indicates the state of the storage volume.

volumeARN :: Maybe Text

The Amazon Resource Name (ARN) for the storage volume. For example, the following is a valid ARN:

arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/volume/vol-1122AABB

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

volumeSizeInBytes :: Maybe Integer

The size of the volume in bytes.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

volumeId :: Maybe Text

The unique identifier assigned to the volume. This ID becomes part of the volume Amazon Resource Name (ARN), which you use as input for other operations.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

gatewayId :: Maybe Text

The unique identifier assigned to your gateway during activation. This ID becomes part of the gateway Amazon Resource Name (ARN), which you use as input for other operations.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

volumeType :: Maybe Text

One of the VolumeType enumeration values describing the type of the volume.

newVolumeInfo :: VolumeInfo

Create a value of VolumeInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:VolumeInfo', volumeInfo_gatewayARN - Undocumented member.

$sel:volumeAttachmentStatus:VolumeInfo', volumeInfo_volumeAttachmentStatus - One of the VolumeStatus values that indicates the state of the storage volume.

$sel:volumeARN:VolumeInfo', volumeInfo_volumeARN - The Amazon Resource Name (ARN) for the storage volume. For example, the following is a valid ARN:

arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/volume/vol-1122AABB

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

$sel:volumeSizeInBytes:VolumeInfo', volumeInfo_volumeSizeInBytes - The size of the volume in bytes.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

$sel:volumeId:VolumeInfo', volumeInfo_volumeId - The unique identifier assigned to the volume. This ID becomes part of the volume Amazon Resource Name (ARN), which you use as input for other operations.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

$sel:gatewayId:VolumeInfo', volumeInfo_gatewayId - The unique identifier assigned to your gateway during activation. This ID becomes part of the gateway Amazon Resource Name (ARN), which you use as input for other operations.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

$sel:volumeType:VolumeInfo', volumeInfo_volumeType - One of the VolumeType enumeration values describing the type of the volume.

volumeInfo_gatewayARN :: Lens' VolumeInfo (Maybe Text)

Undocumented member.

volumeInfo_volumeAttachmentStatus :: Lens' VolumeInfo (Maybe Text)

One of the VolumeStatus values that indicates the state of the storage volume.

volumeInfo_volumeARN :: Lens' VolumeInfo (Maybe Text)

The Amazon Resource Name (ARN) for the storage volume. For example, the following is a valid ARN:

arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/volume/vol-1122AABB

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

volumeInfo_volumeSizeInBytes :: Lens' VolumeInfo (Maybe Integer)

The size of the volume in bytes.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

volumeInfo_volumeId :: Lens' VolumeInfo (Maybe Text)

The unique identifier assigned to the volume. This ID becomes part of the volume Amazon Resource Name (ARN), which you use as input for other operations.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

volumeInfo_gatewayId :: Lens' VolumeInfo (Maybe Text)

The unique identifier assigned to your gateway during activation. This ID becomes part of the gateway Amazon Resource Name (ARN), which you use as input for other operations.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

volumeInfo_volumeType :: Lens' VolumeInfo (Maybe Text)

One of the VolumeType enumeration values describing the type of the volume.