Amazonka.StorageGateway.UpdateBandwidthRateLimitSchedule

Description

Updates the bandwidth rate limit schedule for a specified gateway. By default, gateways do not have bandwidth rate limit schedules, which means no bandwidth rate limiting is in effect. Use this to initiate or update a gateway's bandwidth rate limit schedule. This operation is supported in the volume and tape gateway types.

Creating a Request

data UpdateBandwidthRateLimitSchedule

See: newUpdateBandwidthRateLimitSchedule smart constructor.

Constructors

UpdateBandwidthRateLimitSchedule'

Fields

gatewayARN :: Text

bandwidthRateLimitIntervals :: [BandwidthRateLimitInterval]

An array containing bandwidth rate limit schedule intervals for a gateway. When no bandwidth rate limit intervals have been scheduled, the array is empty.

newUpdateBandwidthRateLimitSchedule

Arguments

:: Text
$sel:gatewayARN:UpdateBandwidthRateLimitSchedule'

-> UpdateBandwidthRateLimitSchedule

Create a value of UpdateBandwidthRateLimitSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:UpdateBandwidthRateLimitSchedule', updateBandwidthRateLimitSchedule_gatewayARN - Undocumented member.

$sel:bandwidthRateLimitIntervals:UpdateBandwidthRateLimitSchedule', updateBandwidthRateLimitSchedule_bandwidthRateLimitIntervals - An array containing bandwidth rate limit schedule intervals for a gateway. When no bandwidth rate limit intervals have been scheduled, the array is empty.

Request Lenses

updateBandwidthRateLimitSchedule_gatewayARN :: Lens' UpdateBandwidthRateLimitSchedule Text

Undocumented member.

updateBandwidthRateLimitSchedule_bandwidthRateLimitIntervals :: Lens' UpdateBandwidthRateLimitSchedule [BandwidthRateLimitInterval]

An array containing bandwidth rate limit schedule intervals for a gateway. When no bandwidth rate limit intervals have been scheduled, the array is empty.

Destructuring the Response

data UpdateBandwidthRateLimitScheduleResponse

See: newUpdateBandwidthRateLimitScheduleResponse smart constructor.

Constructors

UpdateBandwidthRateLimitScheduleResponse'

Fields

gatewayARN :: Maybe Text

httpStatus :: Int

The response's http status code.

newUpdateBandwidthRateLimitScheduleResponse

Arguments

:: Int
$sel:httpStatus:UpdateBandwidthRateLimitScheduleResponse'

-> UpdateBandwidthRateLimitScheduleResponse

Create a value of UpdateBandwidthRateLimitScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayARN:UpdateBandwidthRateLimitSchedule', updateBandwidthRateLimitScheduleResponse_gatewayARN - Undocumented member.

$sel:httpStatus:UpdateBandwidthRateLimitScheduleResponse', updateBandwidthRateLimitScheduleResponse_httpStatus - The response's http status code.

Response Lenses

updateBandwidthRateLimitScheduleResponse_gatewayARN :: Lens' UpdateBandwidthRateLimitScheduleResponse (Maybe Text)

Undocumented member.

updateBandwidthRateLimitScheduleResponse_httpStatus :: Lens' UpdateBandwidthRateLimitScheduleResponse Int

The response's http status code.