Amazonka.StorageGateway.UpdateNFSFileShare

Description

Updates a Network File System (NFS) file share. This operation is only supported in S3 File Gateways.

To leave a file share field unchanged, set the corresponding input field to null.

Updates the following file share settings:

Default storage class for your S3 bucket

Metadata defaults for your S3 bucket

Allowed NFS clients for your file share

Squash settings

Write status of your file share

Creating a Request

data UpdateNFSFileShare

UpdateNFSFileShareInput

See: newUpdateNFSFileShare smart constructor.

Constructors

UpdateNFSFileShare'

Fields

auditDestinationARN :: Maybe Text

The Amazon Resource Name (ARN) of the storage used for audit logs.

kmsKey :: Maybe Text

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

cacheAttributes :: Maybe CacheAttributes

Specifies refresh cache information for the file share.

objectACL :: Maybe ObjectACL

A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

kmsEncrypted :: Maybe Bool

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

defaultStorageClass :: Maybe Text

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

fileShareName :: Maybe Text

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN, or if an access point or access point alias is used.

notificationPolicy :: Maybe Text

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

squash :: Maybe Text

The user mapped to anonymous user.

Valid values are the following:

RootSquash: Only root is mapped to anonymous user.

NoSquash: No one is mapped to anonymous user.

AllSquash: Everyone is mapped to anonymous user.

requesterPays :: Maybe Bool

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

nFSFileShareDefaults :: Maybe NFSFileShareDefaults

The default values for the file share. Optional.

clientList :: Maybe (NonEmpty Text)

The list of clients that are allowed to access the S3 File Gateway. The list must contain either valid IP addresses or valid CIDR blocks.

guessMIMETypeEnabled :: Maybe Bool

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

readOnly :: Maybe Bool

A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

fileShareARN :: Text

The Amazon Resource Name (ARN) of the file share to be updated.

newUpdateNFSFileShare

Arguments

:: Text
$sel:fileShareARN:UpdateNFSFileShare'

-> UpdateNFSFileShare

Create a value of UpdateNFSFileShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:auditDestinationARN:UpdateNFSFileShare', updateNFSFileShare_auditDestinationARN - The Amazon Resource Name (ARN) of the storage used for audit logs.

$sel:kmsKey:UpdateNFSFileShare', updateNFSFileShare_kmsKey - The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

$sel:cacheAttributes:UpdateNFSFileShare', updateNFSFileShare_cacheAttributes - Specifies refresh cache information for the file share.

$sel:objectACL:UpdateNFSFileShare', updateNFSFileShare_objectACL - A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

$sel:kmsEncrypted:UpdateNFSFileShare', updateNFSFileShare_kmsEncrypted - Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

$sel:defaultStorageClass:UpdateNFSFileShare', updateNFSFileShare_defaultStorageClass - The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

$sel:fileShareName:UpdateNFSFileShare', updateNFSFileShare_fileShareName - The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN, or if an access point or access point alias is used.

$sel:notificationPolicy:UpdateNFSFileShare', updateNFSFileShare_notificationPolicy - The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

$sel:squash:UpdateNFSFileShare', updateNFSFileShare_squash - The user mapped to anonymous user.

Valid values are the following:

RootSquash: Only root is mapped to anonymous user.

NoSquash: No one is mapped to anonymous user.

AllSquash: Everyone is mapped to anonymous user.

$sel:requesterPays:UpdateNFSFileShare', updateNFSFileShare_requesterPays - A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

$sel:nFSFileShareDefaults:UpdateNFSFileShare', updateNFSFileShare_nFSFileShareDefaults - The default values for the file share. Optional.

$sel:clientList:UpdateNFSFileShare', updateNFSFileShare_clientList - The list of clients that are allowed to access the S3 File Gateway. The list must contain either valid IP addresses or valid CIDR blocks.

$sel:guessMIMETypeEnabled:UpdateNFSFileShare', updateNFSFileShare_guessMIMETypeEnabled - A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

$sel:readOnly:UpdateNFSFileShare', updateNFSFileShare_readOnly - A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

$sel:fileShareARN:UpdateNFSFileShare', updateNFSFileShare_fileShareARN - The Amazon Resource Name (ARN) of the file share to be updated.

Request Lenses

updateNFSFileShare_auditDestinationARN :: Lens' UpdateNFSFileShare (Maybe Text)

The Amazon Resource Name (ARN) of the storage used for audit logs.

updateNFSFileShare_kmsKey :: Lens' UpdateNFSFileShare (Maybe Text)

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) used for Amazon S3 server-side encryption. Storage Gateway does not support asymmetric CMKs. This value can only be set when KMSEncrypted is true. Optional.

updateNFSFileShare_cacheAttributes :: Lens' UpdateNFSFileShare (Maybe CacheAttributes)

Specifies refresh cache information for the file share.

updateNFSFileShare_objectACL :: Lens' UpdateNFSFileShare (Maybe ObjectACL)

A value that sets the access control list (ACL) permission for objects in the S3 bucket that a S3 File Gateway puts objects into. The default value is private.

updateNFSFileShare_kmsEncrypted :: Lens' UpdateNFSFileShare (Maybe Bool)

Set to true to use Amazon S3 server-side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

Valid Values: true | false

updateNFSFileShare_defaultStorageClass :: Lens' UpdateNFSFileShare (Maybe Text)

The default storage class for objects put into an Amazon S3 bucket by the S3 File Gateway. The default value is S3_INTELLIGENT_TIERING. Optional.

Valid Values: S3_STANDARD | S3_INTELLIGENT_TIERING | S3_STANDARD_IA | S3_ONEZONE_IA

updateNFSFileShare_fileShareName :: Lens' UpdateNFSFileShare (Maybe Text)

The name of the file share. Optional.

FileShareName must be set if an S3 prefix name is set in LocationARN, or if an access point or access point alias is used.

updateNFSFileShare_notificationPolicy :: Lens' UpdateNFSFileShare (Maybe Text)

The notification policy of the file share. SettlingTimeInSeconds controls the number of seconds to wait after the last point in time a client wrote to a file before generating an ObjectUploaded notification. Because clients can make many small writes to files, it's best to set this parameter for as long as possible to avoid generating multiple notifications for the same file in a small time period.

SettlingTimeInSeconds has no effect on the timing of the object uploading to Amazon S3, only the timing of the notification.

The following example sets NotificationPolicy on with SettlingTimeInSeconds set to 60.

{\"Upload\": {\"SettlingTimeInSeconds\": 60}}

The following example sets NotificationPolicy off.

{}

updateNFSFileShare_squash :: Lens' UpdateNFSFileShare (Maybe Text)

The user mapped to anonymous user.

Valid values are the following:

RootSquash: Only root is mapped to anonymous user.

NoSquash: No one is mapped to anonymous user.

AllSquash: Everyone is mapped to anonymous user.

updateNFSFileShare_requesterPays :: Lens' UpdateNFSFileShare (Maybe Bool)

A value that sets who pays the cost of the request and the cost associated with data download from the S3 bucket. If this value is set to true, the requester pays the costs; otherwise, the S3 bucket owner pays. However, the S3 bucket owner always pays the cost of storing data.

RequesterPays is a configuration for the S3 bucket that backs the file share, so make sure that the configuration on the file share is the same as the S3 bucket configuration.

Valid Values: true | false

updateNFSFileShare_nFSFileShareDefaults :: Lens' UpdateNFSFileShare (Maybe NFSFileShareDefaults)

The default values for the file share. Optional.

updateNFSFileShare_clientList :: Lens' UpdateNFSFileShare (Maybe (NonEmpty Text))

The list of clients that are allowed to access the S3 File Gateway. The list must contain either valid IP addresses or valid CIDR blocks.

updateNFSFileShare_guessMIMETypeEnabled :: Lens' UpdateNFSFileShare (Maybe Bool)

A value that enables guessing of the MIME type for uploaded objects based on file extensions. Set this value to true to enable MIME type guessing, otherwise set to false. The default value is true.

Valid Values: true | false

updateNFSFileShare_readOnly :: Lens' UpdateNFSFileShare (Maybe Bool)

A value that sets the write status of a file share. Set this value to true to set the write status to read-only, otherwise set to false.

Valid Values: true | false

updateNFSFileShare_fileShareARN :: Lens' UpdateNFSFileShare Text

The Amazon Resource Name (ARN) of the file share to be updated.

Destructuring the Response

data UpdateNFSFileShareResponse

UpdateNFSFileShareOutput

See: newUpdateNFSFileShareResponse smart constructor.

Constructors

UpdateNFSFileShareResponse'

Fields

fileShareARN :: Maybe Text

The Amazon Resource Name (ARN) of the updated file share.

httpStatus :: Int

The response's http status code.

newUpdateNFSFileShareResponse

Arguments

:: Int
$sel:httpStatus:UpdateNFSFileShareResponse'

-> UpdateNFSFileShareResponse

Create a value of UpdateNFSFileShareResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileShareARN:UpdateNFSFileShare', updateNFSFileShareResponse_fileShareARN - The Amazon Resource Name (ARN) of the updated file share.

$sel:httpStatus:UpdateNFSFileShareResponse', updateNFSFileShareResponse_httpStatus - The response's http status code.

Response Lenses

updateNFSFileShareResponse_fileShareARN :: Lens' UpdateNFSFileShareResponse (Maybe Text)

The Amazon Resource Name (ARN) of the updated file share.

updateNFSFileShareResponse_httpStatus :: Lens' UpdateNFSFileShareResponse Int

The response's http status code.