Amazonka.StorageGateway.UpdateSnapshotSchedule

Description

Updates a snapshot schedule configured for a gateway volume. This operation is only supported in the cached volume and stored volume gateway types.

The default snapshot schedule for volume is once every 24 hours, starting at the creation time of the volume. You can use this API to change the snapshot schedule configured for the volume.

In the request you must identify the gateway volume whose snapshot schedule you want to update, and the schedule information, including when you want the snapshot to begin on a day and the frequency (in hours) of snapshots.

Creating a Request

data UpdateSnapshotSchedule

A JSON object containing one or more of the following fields:

UpdateSnapshotScheduleInput$Description

UpdateSnapshotScheduleInput$RecurrenceInHours

UpdateSnapshotScheduleInput$StartAt

UpdateSnapshotScheduleInput$VolumeARN

See: newUpdateSnapshotSchedule smart constructor.

Constructors

UpdateSnapshotSchedule'

Fields

description :: Maybe Text

Optional description of the snapshot that overwrites the existing description.

tags :: Maybe [Tag]

A list of up to 50 tags that can be assigned to a snapshot. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

volumeARN :: Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

startAt :: Natural

The hour of the day at which the snapshot schedule begins represented as hh, where hh is the hour (0 to 23). The hour of the day is in the time zone of the gateway.

recurrenceInHours :: Natural

Frequency of snapshots. Specify the number of hours between snapshots.

newUpdateSnapshotSchedule

Arguments

:: Text
$sel:volumeARN:UpdateSnapshotSchedule'

-> Natural
$sel:startAt:UpdateSnapshotSchedule'

-> Natural
$sel:recurrenceInHours:UpdateSnapshotSchedule'

-> UpdateSnapshotSchedule

Create a value of UpdateSnapshotSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:UpdateSnapshotSchedule', updateSnapshotSchedule_description - Optional description of the snapshot that overwrites the existing description.

$sel:tags:UpdateSnapshotSchedule', updateSnapshotSchedule_tags - A list of up to 50 tags that can be assigned to a snapshot. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

$sel:volumeARN:UpdateSnapshotSchedule', updateSnapshotSchedule_volumeARN - The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

$sel:startAt:UpdateSnapshotSchedule', updateSnapshotSchedule_startAt - The hour of the day at which the snapshot schedule begins represented as hh, where hh is the hour (0 to 23). The hour of the day is in the time zone of the gateway.

$sel:recurrenceInHours:UpdateSnapshotSchedule', updateSnapshotSchedule_recurrenceInHours - Frequency of snapshots. Specify the number of hours between snapshots.

Request Lenses

updateSnapshotSchedule_description :: Lens' UpdateSnapshotSchedule (Maybe Text)

Optional description of the snapshot that overwrites the existing description.

updateSnapshotSchedule_tags :: Lens' UpdateSnapshotSchedule (Maybe [Tag])

A list of up to 50 tags that can be assigned to a snapshot. Each tag is a key-value pair.

Valid characters for key and value are letters, spaces, and numbers representable in UTF-8 format, and the following special characters: + - = . _ : / @. The maximum length of a tag's key is 128 characters, and the maximum length for a tag's value is 256.

updateSnapshotSchedule_volumeARN :: Lens' UpdateSnapshotSchedule Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

updateSnapshotSchedule_startAt :: Lens' UpdateSnapshotSchedule Natural

The hour of the day at which the snapshot schedule begins represented as hh, where hh is the hour (0 to 23). The hour of the day is in the time zone of the gateway.

updateSnapshotSchedule_recurrenceInHours :: Lens' UpdateSnapshotSchedule Natural

Frequency of snapshots. Specify the number of hours between snapshots.

Destructuring the Response

data UpdateSnapshotScheduleResponse

A JSON object containing the Amazon Resource Name (ARN) of the updated storage volume.

See: newUpdateSnapshotScheduleResponse smart constructor.

Constructors

UpdateSnapshotScheduleResponse'

Fields

volumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

httpStatus :: Int

The response's http status code.

newUpdateSnapshotScheduleResponse

Arguments

:: Int
$sel:httpStatus:UpdateSnapshotScheduleResponse'

-> UpdateSnapshotScheduleResponse

Create a value of UpdateSnapshotScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:volumeARN:UpdateSnapshotSchedule', updateSnapshotScheduleResponse_volumeARN - The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

$sel:httpStatus:UpdateSnapshotScheduleResponse', updateSnapshotScheduleResponse_httpStatus - The response's http status code.

Response Lenses

updateSnapshotScheduleResponse_volumeARN :: Lens' UpdateSnapshotScheduleResponse (Maybe Text)

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

updateSnapshotScheduleResponse_httpStatus :: Lens' UpdateSnapshotScheduleResponse Int

The response's http status code.