Amazonka.Support.DescribeAttachment

Description

Returns the attachment that has the specified ID. Attachments can include screenshots, error logs, or other files that describe your issue. Attachment IDs are generated by the case management system when you add an attachment to a case or case communication. Attachment IDs are returned in the AttachmentDetails objects that are returned by the DescribeCommunications operation.

You must have a Business or Enterprise Support plan to use the AWS Support API.

If you call the AWS Support API from an account that does not have a Business or Enterprise Support plan, the SubscriptionRequiredException error message appears. For information about changing your support plan, see AWS Support.

Creating a Request

data DescribeAttachment

See: newDescribeAttachment smart constructor.

Constructors

DescribeAttachment'

Fields

attachmentId :: Text

The ID of the attachment to return. Attachment IDs are returned by the DescribeCommunications operation.

newDescribeAttachment

Arguments

:: Text
$sel:attachmentId:DescribeAttachment'

-> DescribeAttachment

Create a value of DescribeAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attachmentId:DescribeAttachment', describeAttachment_attachmentId - The ID of the attachment to return. Attachment IDs are returned by the DescribeCommunications operation.

Request Lenses

describeAttachment_attachmentId :: Lens' DescribeAttachment Text

The ID of the attachment to return. Attachment IDs are returned by the DescribeCommunications operation.

Destructuring the Response

data DescribeAttachmentResponse

The content and file name of the attachment returned by the DescribeAttachment operation.

See: newDescribeAttachmentResponse smart constructor.

Constructors

DescribeAttachmentResponse'

Fields

attachment :: Maybe Attachment

This object includes the attachment content and file name.

In the previous response syntax, the value for the data parameter appears as blob, which is represented as a base64-encoded string. The value for fileName is the name of the attachment, such as troubleshoot-screenshot.png.

httpStatus :: Int

The response's http status code.

newDescribeAttachmentResponse

Arguments

:: Int
$sel:httpStatus:DescribeAttachmentResponse'

-> DescribeAttachmentResponse

Create a value of DescribeAttachmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attachment:DescribeAttachmentResponse', describeAttachmentResponse_attachment - This object includes the attachment content and file name.

In the previous response syntax, the value for the data parameter appears as blob, which is represented as a base64-encoded string. The value for fileName is the name of the attachment, such as troubleshoot-screenshot.png.

$sel:httpStatus:DescribeAttachmentResponse', describeAttachmentResponse_httpStatus - The response's http status code.

Response Lenses

describeAttachmentResponse_attachment :: Lens' DescribeAttachmentResponse (Maybe Attachment)

This object includes the attachment content and file name.

In the previous response syntax, the value for the data parameter appears as blob, which is represented as a base64-encoded string. The value for fileName is the name of the attachment, such as troubleshoot-screenshot.png.

describeAttachmentResponse_httpStatus :: Lens' DescribeAttachmentResponse Int

The response's http status code.