Amazonka.Support.DescribeCases

Description

Returns a list of cases that you specify by passing one or more case IDs. You can use the afterTime and beforeTime parameters to filter the cases by date. You can set values for the includeResolvedCases and includeCommunications parameters to specify how much information to return.

The response returns the following in JSON format:

One or more CaseDetails data types.

One or more nextToken values, which specify where to paginate the returned records represented by the CaseDetails objects.

Case data is available for 12 months after creation. If a case was created more than 12 months ago, a request might return an error.

You must have a Business or Enterprise Support plan to use the AWS Support API.

If you call the AWS Support API from an account that does not have a Business or Enterprise Support plan, the SubscriptionRequiredException error message appears. For information about changing your support plan, see AWS Support.

This operation returns paginated results.

Creating a Request

data DescribeCases

See: newDescribeCases smart constructor.

Constructors

DescribeCases'

Fields

includeResolvedCases :: Maybe Bool

Specifies whether to include resolved support cases in the DescribeCases response. By default, resolved cases aren't included.

caseIdList :: Maybe [Text]

A list of ID numbers of the support cases you want returned. The maximum number of cases is 100.

afterTime :: Maybe Text

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

beforeTime :: Maybe Text

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

nextToken :: Maybe Text

A resumption point for pagination.

includeCommunications :: Maybe Bool

Specifies whether to include communications in the DescribeCases response. By default, communications are included.

displayId :: Maybe Text

The ID displayed for a case in the AWS Support Center user interface.

language :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

maxResults :: Maybe Natural

The maximum number of results to return before paginating.

newDescribeCases :: DescribeCases

Create a value of DescribeCases with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeResolvedCases:DescribeCases', describeCases_includeResolvedCases - Specifies whether to include resolved support cases in the DescribeCases response. By default, resolved cases aren't included.

$sel:caseIdList:DescribeCases', describeCases_caseIdList - A list of ID numbers of the support cases you want returned. The maximum number of cases is 100.

$sel:afterTime:DescribeCases', describeCases_afterTime - The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

$sel:beforeTime:DescribeCases', describeCases_beforeTime - The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

$sel:nextToken:DescribeCases', describeCases_nextToken - A resumption point for pagination.

$sel:includeCommunications:DescribeCases', describeCases_includeCommunications - Specifies whether to include communications in the DescribeCases response. By default, communications are included.

$sel:displayId:DescribeCases', describeCases_displayId - The ID displayed for a case in the AWS Support Center user interface.

$sel:language:DescribeCases', describeCases_language - The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

$sel:maxResults:DescribeCases', describeCases_maxResults - The maximum number of results to return before paginating.

Request Lenses

describeCases_includeResolvedCases :: Lens' DescribeCases (Maybe Bool)

Specifies whether to include resolved support cases in the DescribeCases response. By default, resolved cases aren't included.

describeCases_caseIdList :: Lens' DescribeCases (Maybe [Text])

A list of ID numbers of the support cases you want returned. The maximum number of cases is 100.

describeCases_afterTime :: Lens' DescribeCases (Maybe Text)

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

describeCases_beforeTime :: Lens' DescribeCases (Maybe Text)

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

describeCases_nextToken :: Lens' DescribeCases (Maybe Text)

A resumption point for pagination.

describeCases_includeCommunications :: Lens' DescribeCases (Maybe Bool)

Specifies whether to include communications in the DescribeCases response. By default, communications are included.

describeCases_displayId :: Lens' DescribeCases (Maybe Text)

The ID displayed for a case in the AWS Support Center user interface.

describeCases_language :: Lens' DescribeCases (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

describeCases_maxResults :: Lens' DescribeCases (Maybe Natural)

The maximum number of results to return before paginating.

Destructuring the Response

data DescribeCasesResponse

Returns an array of CaseDetails objects and a nextToken that defines a point for pagination in the result set.

See: newDescribeCasesResponse smart constructor.

Constructors

DescribeCasesResponse'

Fields

cases :: Maybe [CaseDetails]

The details for the cases that match the request.

nextToken :: Maybe Text

A resumption point for pagination.

httpStatus :: Int

The response's http status code.

newDescribeCasesResponse

Arguments

:: Int
$sel:httpStatus:DescribeCasesResponse'

-> DescribeCasesResponse

Create a value of DescribeCasesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cases:DescribeCasesResponse', describeCasesResponse_cases - The details for the cases that match the request.

$sel:nextToken:DescribeCases', describeCasesResponse_nextToken - A resumption point for pagination.

$sel:httpStatus:DescribeCasesResponse', describeCasesResponse_httpStatus - The response's http status code.

Response Lenses

describeCasesResponse_cases :: Lens' DescribeCasesResponse (Maybe [CaseDetails])

The details for the cases that match the request.

describeCasesResponse_nextToken :: Lens' DescribeCasesResponse (Maybe Text)

A resumption point for pagination.

describeCasesResponse_httpStatus :: Lens' DescribeCasesResponse Int

The response's http status code.