Amazonka.Support.DescribeCommunications

Description

Returns communications and attachments for one or more support cases. Use the afterTime and beforeTime parameters to filter by date. You can use the caseId parameter to restrict the results to a specific case.

Case data is available for 12 months after creation. If a case was created more than 12 months ago, a request for data might cause an error.

You can use the maxResults and nextToken parameters to control the pagination of the results. Set maxResults to the number of cases that you want to display on each page, and use nextToken to specify the resumption of pagination.

You must have a Business or Enterprise Support plan to use the AWS Support API.

If you call the AWS Support API from an account that does not have a Business or Enterprise Support plan, the SubscriptionRequiredException error message appears. For information about changing your support plan, see AWS Support.

This operation returns paginated results.

Creating a Request

data DescribeCommunications

See: newDescribeCommunications smart constructor.

Constructors

DescribeCommunications'

Fields

afterTime :: Maybe Text

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

beforeTime :: Maybe Text

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

nextToken :: Maybe Text

A resumption point for pagination.

maxResults :: Maybe Natural

The maximum number of results to return before paginating.

caseId :: Text

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

newDescribeCommunications

Arguments

:: Text
$sel:caseId:DescribeCommunications'

-> DescribeCommunications

Create a value of DescribeCommunications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:afterTime:DescribeCommunications', describeCommunications_afterTime - The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

$sel:beforeTime:DescribeCommunications', describeCommunications_beforeTime - The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

$sel:nextToken:DescribeCommunications', describeCommunications_nextToken - A resumption point for pagination.

$sel:maxResults:DescribeCommunications', describeCommunications_maxResults - The maximum number of results to return before paginating.

$sel:caseId:DescribeCommunications', describeCommunications_caseId - The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

Request Lenses

describeCommunications_afterTime :: Lens' DescribeCommunications (Maybe Text)

The start date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

describeCommunications_beforeTime :: Lens' DescribeCommunications (Maybe Text)

The end date for a filtered date search on support case communications. Case communications are available for 12 months after creation.

describeCommunications_nextToken :: Lens' DescribeCommunications (Maybe Text)

A resumption point for pagination.

describeCommunications_maxResults :: Lens' DescribeCommunications (Maybe Natural)

The maximum number of results to return before paginating.

describeCommunications_caseId :: Lens' DescribeCommunications Text

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

Destructuring the Response

data DescribeCommunicationsResponse

The communications returned by the DescribeCommunications operation.

See: newDescribeCommunicationsResponse smart constructor.

Constructors

DescribeCommunicationsResponse'

Fields

nextToken :: Maybe Text

A resumption point for pagination.

communications :: Maybe [Communication]

The communications for the case.

httpStatus :: Int

The response's http status code.

newDescribeCommunicationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeCommunicationsResponse'

-> DescribeCommunicationsResponse

Create a value of DescribeCommunicationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeCommunications', describeCommunicationsResponse_nextToken - A resumption point for pagination.

$sel:communications:DescribeCommunicationsResponse', describeCommunicationsResponse_communications - The communications for the case.

$sel:httpStatus:DescribeCommunicationsResponse', describeCommunicationsResponse_httpStatus - The response's http status code.

Response Lenses

describeCommunicationsResponse_nextToken :: Lens' DescribeCommunicationsResponse (Maybe Text)

A resumption point for pagination.

describeCommunicationsResponse_communications :: Lens' DescribeCommunicationsResponse (Maybe [Communication])

The communications for the case.

describeCommunicationsResponse_httpStatus :: Lens' DescribeCommunicationsResponse Int

The response's http status code.