Amazonka.Support.DescribeSeverityLevels

Description

Returns the list of severity levels that you can assign to a support case. The severity level for a case is also a field in the CaseDetails data type that you include for a CreateCase request.

You must have a Business or Enterprise Support plan to use the AWS Support API.

If you call the AWS Support API from an account that does not have a Business or Enterprise Support plan, the SubscriptionRequiredException error message appears. For information about changing your support plan, see AWS Support.

Creating a Request

data DescribeSeverityLevels

See: newDescribeSeverityLevels smart constructor.

Constructors

DescribeSeverityLevels'

Fields

language :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

newDescribeSeverityLevels :: DescribeSeverityLevels

Create a value of DescribeSeverityLevels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:language:DescribeSeverityLevels', describeSeverityLevels_language - The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Request Lenses

describeSeverityLevels_language :: Lens' DescribeSeverityLevels (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

Destructuring the Response

data DescribeSeverityLevelsResponse

The list of severity levels returned by the DescribeSeverityLevels operation.

See: newDescribeSeverityLevelsResponse smart constructor.

Constructors

DescribeSeverityLevelsResponse'

Fields

severityLevels :: Maybe [SeverityLevel]

The available severity levels for the support case. Available severity levels are defined by your service level agreement with AWS.

httpStatus :: Int

The response's http status code.

newDescribeSeverityLevelsResponse

Arguments

:: Int
$sel:httpStatus:DescribeSeverityLevelsResponse'

-> DescribeSeverityLevelsResponse

Create a value of DescribeSeverityLevelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:severityLevels:DescribeSeverityLevelsResponse', describeSeverityLevelsResponse_severityLevels - The available severity levels for the support case. Available severity levels are defined by your service level agreement with AWS.

$sel:httpStatus:DescribeSeverityLevelsResponse', describeSeverityLevelsResponse_httpStatus - The response's http status code.

Response Lenses

describeSeverityLevelsResponse_severityLevels :: Lens' DescribeSeverityLevelsResponse (Maybe [SeverityLevel])

The available severity levels for the support case. Available severity levels are defined by your service level agreement with AWS.

describeSeverityLevelsResponse_httpStatus :: Lens' DescribeSeverityLevelsResponse Int

The response's http status code.