Amazonka.Support.DescribeTrustedAdvisorCheckSummaries

Description

Returns the results for the AWS Trusted Advisor check summaries for the check IDs that you specified. You can get the check IDs by calling the DescribeTrustedAdvisorChecks operation.

The response contains an array of TrustedAdvisorCheckSummary objects.

You must have a Business or Enterprise Support plan to use the AWS Support API.

If you call the AWS Support API from an account that does not have a Business or Enterprise Support plan, the SubscriptionRequiredException error message appears. For information about changing your support plan, see AWS Support.

Creating a Request

data DescribeTrustedAdvisorCheckSummaries

See: newDescribeTrustedAdvisorCheckSummaries smart constructor.

Constructors

DescribeTrustedAdvisorCheckSummaries'

Fields

checkIds :: [Text]

The IDs of the Trusted Advisor checks.

newDescribeTrustedAdvisorCheckSummaries :: DescribeTrustedAdvisorCheckSummaries

Create a value of DescribeTrustedAdvisorCheckSummaries with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkIds:DescribeTrustedAdvisorCheckSummaries', describeTrustedAdvisorCheckSummaries_checkIds - The IDs of the Trusted Advisor checks.

Request Lenses

describeTrustedAdvisorCheckSummaries_checkIds :: Lens' DescribeTrustedAdvisorCheckSummaries [Text]

The IDs of the Trusted Advisor checks.

Destructuring the Response

data DescribeTrustedAdvisorCheckSummariesResponse

The summaries of the Trusted Advisor checks returned by the DescribeTrustedAdvisorCheckSummaries operation.

See: newDescribeTrustedAdvisorCheckSummariesResponse smart constructor.

Constructors

DescribeTrustedAdvisorCheckSummariesResponse'

Fields

httpStatus :: Int

The response's http status code.

summaries :: [TrustedAdvisorCheckSummary]

The summary information for the requested Trusted Advisor checks.

newDescribeTrustedAdvisorCheckSummariesResponse

Arguments

:: Int
$sel:httpStatus:DescribeTrustedAdvisorCheckSummariesResponse'

-> DescribeTrustedAdvisorCheckSummariesResponse

Create a value of DescribeTrustedAdvisorCheckSummariesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeTrustedAdvisorCheckSummariesResponse', describeTrustedAdvisorCheckSummariesResponse_httpStatus - The response's http status code.

$sel:summaries:DescribeTrustedAdvisorCheckSummariesResponse', describeTrustedAdvisorCheckSummariesResponse_summaries - The summary information for the requested Trusted Advisor checks.

Response Lenses

describeTrustedAdvisorCheckSummariesResponse_httpStatus :: Lens' DescribeTrustedAdvisorCheckSummariesResponse Int

The response's http status code.

describeTrustedAdvisorCheckSummariesResponse_summaries :: Lens' DescribeTrustedAdvisorCheckSummariesResponse [TrustedAdvisorCheckSummary]

The summary information for the requested Trusted Advisor checks.