Amazonka.Support.Types.CaseDetails

Description

Documentation

data CaseDetails

A JSON-formatted object that contains the metadata for a support case. It is contained in the response from a DescribeCases request. CaseDetails contains the following fields:

caseId - The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

categoryCode - The category of problem for the support case. Corresponds to the CategoryCode values returned by a call to DescribeServices.

displayId - The identifier for the case on pages in the AWS Support Center.

language - The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

nextToken - A resumption point for pagination.

recentCommunications - One or more Communication objects. Fields of these objects are attachments, body, caseId, submittedBy, and timeCreated.

serviceCode - The identifier for the AWS service that corresponds to the service code defined in the call to DescribeServices.

severityCode - The severity code assigned to the case. Contains one of the values returned by the call to DescribeSeverityLevels. The possible values are: low, normal, high, urgent, and critical.

status - The status of the case in the AWS Support Center. Valid values:

opened

pending-customer-action

reopened

resolved

unassigned

work-in-progress

subject - The subject line of the case.

submittedBy - The email address of the account that submitted the case.

timeCreated - The time the case was created, in ISO-8601 format.

See: newCaseDetails smart constructor.

Constructors

CaseDetails'

Fields

subject :: Maybe Text

The subject line for the case in the AWS Support Center.

status :: Maybe Text

The status of the case.

Valid values:

opened

pending-customer-action

reopened

resolved

unassigned

work-in-progress

recentCommunications :: Maybe RecentCaseCommunications

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

severityCode :: Maybe Text

The code for the severity level returned by the call to DescribeSeverityLevels.

caseId :: Maybe Text

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

ccEmailAddresses :: Maybe [Text]

The email addresses that receive copies of communication about the case.

displayId :: Maybe Text

The ID displayed for the case in the AWS Support Center. This is a numeric string.

submittedBy :: Maybe Text

The email address of the account that submitted the case.

language :: Maybe Text

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

timeCreated :: Maybe Text

The time that the case was created in the AWS Support Center.

categoryCode :: Maybe Text

The category of problem for the support case.

serviceCode :: Maybe Text

The code for the AWS service. You can get a list of codes and the corresponding service names by calling DescribeServices.

newCaseDetails :: CaseDetails

Create a value of CaseDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subject:CaseDetails', caseDetails_subject - The subject line for the case in the AWS Support Center.

$sel:status:CaseDetails', caseDetails_status - The status of the case.

Valid values:

opened

pending-customer-action

reopened

resolved

unassigned

work-in-progress

$sel:recentCommunications:CaseDetails', caseDetails_recentCommunications - The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

$sel:severityCode:CaseDetails', caseDetails_severityCode - The code for the severity level returned by the call to DescribeSeverityLevels.

$sel:caseId:CaseDetails', caseDetails_caseId - The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

$sel:ccEmailAddresses:CaseDetails', caseDetails_ccEmailAddresses - The email addresses that receive copies of communication about the case.

$sel:displayId:CaseDetails', caseDetails_displayId - The ID displayed for the case in the AWS Support Center. This is a numeric string.

$sel:submittedBy:CaseDetails', caseDetails_submittedBy - The email address of the account that submitted the case.

$sel:language:CaseDetails', caseDetails_language - The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

$sel:timeCreated:CaseDetails', caseDetails_timeCreated - The time that the case was created in the AWS Support Center.

$sel:categoryCode:CaseDetails', caseDetails_categoryCode - The category of problem for the support case.

$sel:serviceCode:CaseDetails', caseDetails_serviceCode - The code for the AWS service. You can get a list of codes and the corresponding service names by calling DescribeServices.

caseDetails_subject :: Lens' CaseDetails (Maybe Text)

The subject line for the case in the AWS Support Center.

caseDetails_status :: Lens' CaseDetails (Maybe Text)

The status of the case.

Valid values:

opened

pending-customer-action

reopened

resolved

unassigned

work-in-progress

caseDetails_recentCommunications :: Lens' CaseDetails (Maybe RecentCaseCommunications)

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

caseDetails_severityCode :: Lens' CaseDetails (Maybe Text)

The code for the severity level returned by the call to DescribeSeverityLevels.

caseDetails_caseId :: Lens' CaseDetails (Maybe Text)

The support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

caseDetails_ccEmailAddresses :: Lens' CaseDetails (Maybe [Text])

The email addresses that receive copies of communication about the case.

caseDetails_displayId :: Lens' CaseDetails (Maybe Text)

The ID displayed for the case in the AWS Support Center. This is a numeric string.

caseDetails_submittedBy :: Lens' CaseDetails (Maybe Text)

The email address of the account that submitted the case.

caseDetails_language :: Lens' CaseDetails (Maybe Text)

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

caseDetails_timeCreated :: Lens' CaseDetails (Maybe Text)

The time that the case was created in the AWS Support Center.

caseDetails_categoryCode :: Lens' CaseDetails (Maybe Text)

The category of problem for the support case.

caseDetails_serviceCode :: Lens' CaseDetails (Maybe Text)

The code for the AWS service. You can get a list of codes and the corresponding service names by calling DescribeServices.