Amazonka.Support.Types.TrustedAdvisorCheckResult

Description

Documentation

data TrustedAdvisorCheckResult

The results of a Trusted Advisor check returned by DescribeTrustedAdvisorCheckResult.

See: newTrustedAdvisorCheckResult smart constructor.

Constructors

TrustedAdvisorCheckResult'

Fields

checkId :: Text

The unique identifier for the Trusted Advisor check.

timestamp :: Text

The time of the last refresh of the check.

status :: Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

resourcesSummary :: TrustedAdvisorResourcesSummary

categorySpecificSummary :: TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

flaggedResources :: [TrustedAdvisorResourceDetail]

The details about each resource listed in the check result.

newTrustedAdvisorCheckResult

Arguments

:: Text
$sel:checkId:TrustedAdvisorCheckResult'

-> Text
$sel:timestamp:TrustedAdvisorCheckResult'

-> Text
$sel:status:TrustedAdvisorCheckResult'

-> TrustedAdvisorResourcesSummary
$sel:resourcesSummary:TrustedAdvisorCheckResult'

-> TrustedAdvisorCategorySpecificSummary
$sel:categorySpecificSummary:TrustedAdvisorCheckResult'

-> TrustedAdvisorCheckResult

Create a value of TrustedAdvisorCheckResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkId:TrustedAdvisorCheckResult', trustedAdvisorCheckResult_checkId - The unique identifier for the Trusted Advisor check.

$sel:timestamp:TrustedAdvisorCheckResult', trustedAdvisorCheckResult_timestamp - The time of the last refresh of the check.

$sel:status:TrustedAdvisorCheckResult', trustedAdvisorCheckResult_status - The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

$sel:resourcesSummary:TrustedAdvisorCheckResult', trustedAdvisorCheckResult_resourcesSummary - Undocumented member.

$sel:categorySpecificSummary:TrustedAdvisorCheckResult', trustedAdvisorCheckResult_categorySpecificSummary - Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

$sel:flaggedResources:TrustedAdvisorCheckResult', trustedAdvisorCheckResult_flaggedResources - The details about each resource listed in the check result.

trustedAdvisorCheckResult_checkId :: Lens' TrustedAdvisorCheckResult Text

The unique identifier for the Trusted Advisor check.

trustedAdvisorCheckResult_timestamp :: Lens' TrustedAdvisorCheckResult Text

The time of the last refresh of the check.

trustedAdvisorCheckResult_status :: Lens' TrustedAdvisorCheckResult Text

The alert status of the check: "ok" (green), "warning" (yellow), "error" (red), or "not_available".

trustedAdvisorCheckResult_resourcesSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorResourcesSummary

Undocumented member.

trustedAdvisorCheckResult_categorySpecificSummary :: Lens' TrustedAdvisorCheckResult TrustedAdvisorCategorySpecificSummary

Summary information that relates to the category of the check. Cost Optimizing is the only category that is currently supported.

trustedAdvisorCheckResult_flaggedResources :: Lens' TrustedAdvisorCheckResult [TrustedAdvisorResourceDetail]

The details about each resource listed in the check result.