Amazonka.Support.Types.TrustedAdvisorResourceDetail

Description

Documentation

data TrustedAdvisorResourceDetail

Contains information about a resource identified by a Trusted Advisor check.

See: newTrustedAdvisorResourceDetail smart constructor.

Constructors

TrustedAdvisorResourceDetail'

Fields

isSuppressed :: Maybe Bool

Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

region :: Maybe Text

The AWS Region in which the identified resource is located.

status :: Text

The status code for the resource identified in the Trusted Advisor check.

resourceId :: Text

The unique identifier for the identified resource.

metadata :: [Text]

Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

newTrustedAdvisorResourceDetail

Arguments

:: Text
$sel:status:TrustedAdvisorResourceDetail'

-> Text
$sel:resourceId:TrustedAdvisorResourceDetail'

-> TrustedAdvisorResourceDetail

Create a value of TrustedAdvisorResourceDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isSuppressed:TrustedAdvisorResourceDetail', trustedAdvisorResourceDetail_isSuppressed - Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

$sel:region:TrustedAdvisorResourceDetail', trustedAdvisorResourceDetail_region - The AWS Region in which the identified resource is located.

$sel:status:TrustedAdvisorResourceDetail', trustedAdvisorResourceDetail_status - The status code for the resource identified in the Trusted Advisor check.

$sel:resourceId:TrustedAdvisorResourceDetail', trustedAdvisorResourceDetail_resourceId - The unique identifier for the identified resource.

$sel:metadata:TrustedAdvisorResourceDetail', trustedAdvisorResourceDetail_metadata - Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.

trustedAdvisorResourceDetail_isSuppressed :: Lens' TrustedAdvisorResourceDetail (Maybe Bool)

Specifies whether the AWS resource was ignored by Trusted Advisor because it was marked as suppressed by the user.

trustedAdvisorResourceDetail_region :: Lens' TrustedAdvisorResourceDetail (Maybe Text)

The AWS Region in which the identified resource is located.

trustedAdvisorResourceDetail_status :: Lens' TrustedAdvisorResourceDetail Text

The status code for the resource identified in the Trusted Advisor check.

trustedAdvisorResourceDetail_resourceId :: Lens' TrustedAdvisorResourceDetail Text

The unique identifier for the identified resource.

trustedAdvisorResourceDetail_metadata :: Lens' TrustedAdvisorResourceDetail [Text]

Additional information about the identified resource. The exact metadata and its order can be obtained by inspecting the TrustedAdvisorCheckDescription object returned by the call to DescribeTrustedAdvisorChecks. Metadata contains all the data that is shown in the Excel download, even in those cases where the UI shows just summary data.