Amazonka.Synthetics

Description

Derived from API version 2017-10-11 of the AWS service descriptions, licensed under Apache 2.0.

Amazon CloudWatch Synthetics

You can use Amazon CloudWatch Synthetics to continually monitor your services. You can create and manage canaries, which are modular, lightweight scripts that monitor your endpoints and APIs from the outside-in. You can set up your canaries to run 24 hours a day, once per minute. The canaries help you check the availability and latency of your web services and troubleshoot anomalies by investigating load time data, screenshots of the UI, logs, and metrics. The canaries seamlessly integrate with CloudWatch ServiceLens to help you trace the causes of impacted nodes in your applications. For more information, see Using ServiceLens to Monitor the Health of Your Applications in the Amazon CloudWatch User Guide.

Before you create and manage canaries, be aware of the security considerations. For more information, see Security Considerations for Synthetics Canaries.

Service Configuration

defaultService :: Service

API version 2017-10-11 of the Amazon Synthetics SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Synthetics.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

A parameter could not be validated.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

A conflicting operation is already in progress.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An unknown internal error occurred.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One of the specified resources was not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

UpdateCanary

data UpdateCanary

See: newUpdateCanary smart constructor.

Constructors

UpdateCanary' (Maybe Natural) (Maybe CanaryScheduleInput) (Maybe Text) (Maybe CanaryRunConfigInput) (Maybe Text) (Maybe Text) (Maybe Natural) (Maybe ArtifactConfigInput) (Maybe VpcConfigInput) (Maybe VisualReferenceInput) (Maybe CanaryCodeInput) Text

newUpdateCanary

Arguments

:: Text
$sel:name:UpdateCanary'

-> UpdateCanary

Create a value of UpdateCanary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:successRetentionPeriodInDays:UpdateCanary', updateCanary_successRetentionPeriodInDays - The number of days to retain data about successful runs of this canary.

$sel:schedule:UpdateCanary', updateCanary_schedule - A structure that contains information about how often the canary is to run, and when these runs are to stop.

$sel:artifactS3Location:UpdateCanary', updateCanary_artifactS3Location - The location in Amazon S3 where Synthetics stores artifacts from the test runs of this canary. Artifacts include the log file, screenshots, and HAR files. The name of the S3 bucket can't include a period (.).

$sel:runConfig:UpdateCanary', updateCanary_runConfig - A structure that contains the timeout value that is used for each individual run of the canary.

$sel:executionRoleArn:UpdateCanary', updateCanary_executionRoleArn - The ARN of the IAM role to be used to run the canary. This role must already exist, and must include lambda.amazonaws.com as a principal in the trust policy. The role must also have the following permissions:

s3:PutObject

s3:GetBucketLocation

s3:ListAllMyBuckets

cloudwatch:PutMetricData

logs:CreateLogGroup

logs:CreateLogStream

logs:CreateLogStream

$sel:runtimeVersion:UpdateCanary', updateCanary_runtimeVersion - Specifies the runtime version to use for the canary. For a list of valid runtime versions and for more information about runtime versions, see Canary Runtime Versions.

$sel:failureRetentionPeriodInDays:UpdateCanary', updateCanary_failureRetentionPeriodInDays - The number of days to retain data about failed runs of this canary.

$sel:artifactConfig:UpdateCanary', updateCanary_artifactConfig - A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

$sel:vpcConfig:UpdateCanary', updateCanary_vpcConfig - If this canary is to test an endpoint in a VPC, this structure contains information about the subnet and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.

$sel:visualReference:UpdateCanary', updateCanary_visualReference - Defines the screenshots to use as the baseline for comparisons during visual monitoring comparisons during future runs of this canary. If you omit this parameter, no changes are made to any baseline screenshots that the canary might be using already.

Visual monitoring is supported only on canaries running the syn-puppeteer-node-3.2 runtime or later. For more information, see Visual monitoring and Visual monitoring blueprint

$sel:code:UpdateCanary', updateCanary_code - A structure that includes the entry point from which the canary should start running your script. If the script is stored in an S3 bucket, the bucket name, key, and version are also included.

$sel:name:UpdateCanary', updateCanary_name - The name of the canary that you want to update. To find the names of your canaries, use DescribeCanaries.

You cannot change the name of a canary that has already been created.

data UpdateCanaryResponse

See: newUpdateCanaryResponse smart constructor.

Constructors

UpdateCanaryResponse' Int

newUpdateCanaryResponse

Arguments

:: Int
$sel:httpStatus:UpdateCanaryResponse'

-> UpdateCanaryResponse

Create a value of UpdateCanaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateCanaryResponse', updateCanaryResponse_httpStatus - The response's http status code.

DeleteCanary

data DeleteCanary

See: newDeleteCanary smart constructor.

Constructors

DeleteCanary' Text

newDeleteCanary

Arguments

:: Text
$sel:name:DeleteCanary'

-> DeleteCanary

Create a value of DeleteCanary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteCanary', deleteCanary_name - The name of the canary that you want to delete. To find the names of your canaries, use DescribeCanaries.

data DeleteCanaryResponse

See: newDeleteCanaryResponse smart constructor.

Constructors

DeleteCanaryResponse' Int

newDeleteCanaryResponse

Arguments

:: Int
$sel:httpStatus:DeleteCanaryResponse'

-> DeleteCanaryResponse

Create a value of DeleteCanaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteCanaryResponse', deleteCanaryResponse_httpStatus - The response's http status code.

CreateCanary

data CreateCanary

See: newCreateCanary smart constructor.

Constructors

CreateCanary' (Maybe Natural) (Maybe CanaryRunConfigInput) (Maybe Natural) (Maybe ArtifactConfigInput) (Maybe VpcConfigInput) (Maybe (HashMap Text Text)) Text CanaryCodeInput Text Text CanaryScheduleInput Text

newCreateCanary

Arguments

:: Text
$sel:name:CreateCanary'

-> CanaryCodeInput
$sel:code:CreateCanary'

-> Text
$sel:artifactS3Location:CreateCanary'

-> Text
$sel:executionRoleArn:CreateCanary'

-> CanaryScheduleInput
$sel:schedule:CreateCanary'

-> Text
$sel:runtimeVersion:CreateCanary'

-> CreateCanary

Create a value of CreateCanary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:successRetentionPeriodInDays:CreateCanary', createCanary_successRetentionPeriodInDays - The number of days to retain data about successful runs of this canary. If you omit this field, the default of 31 days is used. The valid range is 1 to 455 days.

$sel:runConfig:CreateCanary', createCanary_runConfig - A structure that contains the configuration for individual canary runs, such as timeout value.

$sel:failureRetentionPeriodInDays:CreateCanary', createCanary_failureRetentionPeriodInDays - The number of days to retain data about failed runs of this canary. If you omit this field, the default of 31 days is used. The valid range is 1 to 455 days.

$sel:artifactConfig:CreateCanary', createCanary_artifactConfig - A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

$sel:vpcConfig:CreateCanary', createCanary_vpcConfig - If this canary is to test an endpoint in a VPC, this structure contains information about the subnet and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.

$sel:tags:CreateCanary', createCanary_tags - A list of key-value pairs to associate with the canary. You can associate as many as 50 tags with a canary.

Tags can help you organize and categorize your resources. You can also use them to scope user permissions, by granting a user permission to access or change only the resources that have certain tag values.

$sel:name:CreateCanary', createCanary_name - The name for this canary. Be sure to give it a descriptive name that distinguishes it from other canaries in your account.

Do not include secrets or proprietary information in your canary names. The canary name makes up part of the canary ARN, and the ARN is included in outbound calls over the internet. For more information, see Security Considerations for Synthetics Canaries.

$sel:code:CreateCanary', createCanary_code - A structure that includes the entry point from which the canary should start running your script. If the script is stored in an S3 bucket, the bucket name, key, and version are also included.

$sel:artifactS3Location:CreateCanary', createCanary_artifactS3Location - The location in Amazon S3 where Synthetics stores artifacts from the test runs of this canary. Artifacts include the log file, screenshots, and HAR files. The name of the S3 bucket can't include a period (.).

$sel:executionRoleArn:CreateCanary', createCanary_executionRoleArn - The ARN of the IAM role to be used to run the canary. This role must already exist, and must include lambda.amazonaws.com as a principal in the trust policy. The role must also have the following permissions:

s3:PutObject

s3:GetBucketLocation

s3:ListAllMyBuckets

cloudwatch:PutMetricData

logs:CreateLogGroup

logs:CreateLogStream

logs:PutLogEvents

$sel:schedule:CreateCanary', createCanary_schedule - A structure that contains information about how often the canary is to run and when these test runs are to stop.

$sel:runtimeVersion:CreateCanary', createCanary_runtimeVersion - Specifies the runtime version to use for the canary. For a list of valid runtime versions and more information about runtime versions, see Canary Runtime Versions.

data CreateCanaryResponse

See: newCreateCanaryResponse smart constructor.

Constructors

CreateCanaryResponse' (Maybe Canary) Int

newCreateCanaryResponse

Arguments

:: Int
$sel:httpStatus:CreateCanaryResponse'

-> CreateCanaryResponse

Create a value of CreateCanaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:canary:CreateCanaryResponse', createCanaryResponse_canary - The full details about the canary you have created.

$sel:httpStatus:CreateCanaryResponse', createCanaryResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The ARN of the canary that you want to view tags for.

The ARN format of a canary is arn:aws:synthetics:Region:account-id:canary:canary-name .

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The list of tag keys and values associated with the canary that you specified.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

GetCanaryRuns

data GetCanaryRuns

See: newGetCanaryRuns smart constructor.

Constructors

GetCanaryRuns' (Maybe Text) (Maybe Natural) Text

newGetCanaryRuns

Arguments

:: Text
$sel:name:GetCanaryRuns'

-> GetCanaryRuns

Create a value of GetCanaryRuns with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetCanaryRuns', getCanaryRuns_nextToken - A token that indicates that there is more data available. You can use this token in a subsequent GetCanaryRuns operation to retrieve the next set of results.

$sel:maxResults:GetCanaryRuns', getCanaryRuns_maxResults - Specify this parameter to limit how many runs are returned each time you use the GetCanaryRuns operation. If you omit this parameter, the default of 100 is used.

$sel:name:GetCanaryRuns', getCanaryRuns_name - The name of the canary that you want to see runs for.

data GetCanaryRunsResponse

See: newGetCanaryRunsResponse smart constructor.

Constructors

GetCanaryRunsResponse' (Maybe Text) (Maybe [CanaryRun]) Int

newGetCanaryRunsResponse

Arguments

:: Int
$sel:httpStatus:GetCanaryRunsResponse'

-> GetCanaryRunsResponse

Create a value of GetCanaryRunsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetCanaryRuns', getCanaryRunsResponse_nextToken - A token that indicates that there is more data available. You can use this token in a subsequent GetCanaryRuns operation to retrieve the next set of results.

$sel:canaryRuns:GetCanaryRunsResponse', getCanaryRunsResponse_canaryRuns - An array of structures. Each structure contains the details of one of the retrieved canary runs.

$sel:httpStatus:GetCanaryRunsResponse', getCanaryRunsResponse_httpStatus - The response's http status code.

GetCanary

data GetCanary

See: newGetCanary smart constructor.

Constructors

GetCanary' Text

newGetCanary

Arguments

:: Text
$sel:name:GetCanary'

-> GetCanary

Create a value of GetCanary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetCanary', getCanary_name - The name of the canary that you want details for.

data GetCanaryResponse

See: newGetCanaryResponse smart constructor.

Constructors

GetCanaryResponse' (Maybe Canary) Int

newGetCanaryResponse

Arguments

:: Int
$sel:httpStatus:GetCanaryResponse'

-> GetCanaryResponse

Create a value of GetCanaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:canary:GetCanaryResponse', getCanaryResponse_canary - A strucure that contains the full information about the canary.

$sel:httpStatus:GetCanaryResponse', getCanaryResponse_httpStatus - The response's http status code.

DescribeRuntimeVersions

data DescribeRuntimeVersions

See: newDescribeRuntimeVersions smart constructor.

Constructors

DescribeRuntimeVersions' (Maybe Text) (Maybe Natural)

newDescribeRuntimeVersions :: DescribeRuntimeVersions

Create a value of DescribeRuntimeVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeRuntimeVersions', describeRuntimeVersions_nextToken - A token that indicates that there is more data available. You can use this token in a subsequent DescribeRuntimeVersions operation to retrieve the next set of results.

$sel:maxResults:DescribeRuntimeVersions', describeRuntimeVersions_maxResults - Specify this parameter to limit how many runs are returned each time you use the DescribeRuntimeVersions operation. If you omit this parameter, the default of 100 is used.

data DescribeRuntimeVersionsResponse

See: newDescribeRuntimeVersionsResponse smart constructor.

Constructors

DescribeRuntimeVersionsResponse' (Maybe [RuntimeVersion]) (Maybe Text) Int

newDescribeRuntimeVersionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeRuntimeVersionsResponse'

-> DescribeRuntimeVersionsResponse

Create a value of DescribeRuntimeVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:runtimeVersions:DescribeRuntimeVersionsResponse', describeRuntimeVersionsResponse_runtimeVersions - An array of objects that display the details about each Synthetics canary runtime version.

$sel:nextToken:DescribeRuntimeVersions', describeRuntimeVersionsResponse_nextToken - A token that indicates that there is more data available. You can use this token in a subsequent DescribeRuntimeVersions operation to retrieve the next set of results.

$sel:httpStatus:DescribeRuntimeVersionsResponse', describeRuntimeVersionsResponse_httpStatus - The response's http status code.

DescribeCanariesLastRun

data DescribeCanariesLastRun

See: newDescribeCanariesLastRun smart constructor.

Constructors

DescribeCanariesLastRun' (Maybe Text) (Maybe Natural)

newDescribeCanariesLastRun :: DescribeCanariesLastRun

Create a value of DescribeCanariesLastRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeCanariesLastRun', describeCanariesLastRun_nextToken - A token that indicates that there is more data available. You can use this token in a subsequent DescribeCanaries operation to retrieve the next set of results.

$sel:maxResults:DescribeCanariesLastRun', describeCanariesLastRun_maxResults - Specify this parameter to limit how many runs are returned each time you use the DescribeLastRun operation. If you omit this parameter, the default of 100 is used.

data DescribeCanariesLastRunResponse

See: newDescribeCanariesLastRunResponse smart constructor.

Constructors

DescribeCanariesLastRunResponse' (Maybe Text) (Maybe [CanaryLastRun]) Int

newDescribeCanariesLastRunResponse

Arguments

:: Int
$sel:httpStatus:DescribeCanariesLastRunResponse'

-> DescribeCanariesLastRunResponse

Create a value of DescribeCanariesLastRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeCanariesLastRun', describeCanariesLastRunResponse_nextToken - A token that indicates that there is more data available. You can use this token in a subsequent DescribeCanariesLastRun operation to retrieve the next set of results.

$sel:canariesLastRun:DescribeCanariesLastRunResponse', describeCanariesLastRunResponse_canariesLastRun - An array that contains the information from the most recent run of each canary.

$sel:httpStatus:DescribeCanariesLastRunResponse', describeCanariesLastRunResponse_httpStatus - The response's http status code.

StartCanary

data StartCanary

See: newStartCanary smart constructor.

Constructors

StartCanary' Text

newStartCanary

Arguments

:: Text
$sel:name:StartCanary'

-> StartCanary

Create a value of StartCanary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:StartCanary', startCanary_name - The name of the canary that you want to run. To find canary names, use DescribeCanaries.

data StartCanaryResponse

See: newStartCanaryResponse smart constructor.

Constructors

StartCanaryResponse' Int

newStartCanaryResponse

Arguments

:: Int
$sel:httpStatus:StartCanaryResponse'

-> StartCanaryResponse

Create a value of StartCanaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartCanaryResponse', startCanaryResponse_httpStatus - The response's http status code.

DescribeCanaries

data DescribeCanaries

See: newDescribeCanaries smart constructor.

Constructors

DescribeCanaries' (Maybe Text) (Maybe Natural)

newDescribeCanaries :: DescribeCanaries

Create a value of DescribeCanaries with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeCanaries', describeCanaries_nextToken - A token that indicates that there is more data available. You can use this token in a subsequent operation to retrieve the next set of results.

$sel:maxResults:DescribeCanaries', describeCanaries_maxResults - Specify this parameter to limit how many canaries are returned each time you use the DescribeCanaries operation. If you omit this parameter, the default of 100 is used.

data DescribeCanariesResponse

See: newDescribeCanariesResponse smart constructor.

Constructors

DescribeCanariesResponse' (Maybe [Canary]) (Maybe Text) Int

newDescribeCanariesResponse

Arguments

:: Int
$sel:httpStatus:DescribeCanariesResponse'

-> DescribeCanariesResponse

Create a value of DescribeCanariesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:canaries:DescribeCanariesResponse', describeCanariesResponse_canaries - Returns an array. Each item in the array contains the full information about one canary.

$sel:nextToken:DescribeCanaries', describeCanariesResponse_nextToken - A token that indicates that there is more data available. You can use this token in a subsequent DescribeCanaries operation to retrieve the next set of results.

$sel:httpStatus:DescribeCanariesResponse', describeCanariesResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The ARN of the canary that you're adding tags to.

The ARN format of a canary is arn:aws:synthetics:Region:account-id:canary:canary-name .

$sel:tags:TagResource', tagResource_tags - The list of key-value pairs to associate with the canary.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text (NonEmpty Text)

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> NonEmpty Text
$sel:tagKeys:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The ARN of the canary that you're removing tags from.

The ARN format of a canary is arn:aws:synthetics:Region:account-id:canary:canary-name .

$sel:tagKeys:UntagResource', untagResource_tagKeys - The list of tag keys to remove from the resource.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

StopCanary

data StopCanary

See: newStopCanary smart constructor.

Constructors

StopCanary' Text

newStopCanary

Arguments

:: Text
$sel:name:StopCanary'

-> StopCanary

Create a value of StopCanary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:StopCanary', stopCanary_name - The name of the canary that you want to stop. To find the names of your canaries, use DescribeCanaries.

data StopCanaryResponse

See: newStopCanaryResponse smart constructor.

Constructors

StopCanaryResponse' Int

newStopCanaryResponse

Arguments

:: Int
$sel:httpStatus:StopCanaryResponse'

-> StopCanaryResponse

Create a value of StopCanaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StopCanaryResponse', stopCanaryResponse_httpStatus - The response's http status code.

Types

CanaryRunState

newtype CanaryRunState

Constructors

CanaryRunState'

Fields

fromCanaryRunState :: Text

Bundled Patterns

pattern CanaryRunState_FAILED :: CanaryRunState

pattern CanaryRunState_PASSED :: CanaryRunState

pattern CanaryRunState_RUNNING :: CanaryRunState

CanaryRunStateReasonCode

newtype CanaryRunStateReasonCode

Constructors

CanaryRunStateReasonCode'

Fields

fromCanaryRunStateReasonCode :: Text

Bundled Patterns

pattern CanaryRunStateReasonCode_CANARY_FAILURE :: CanaryRunStateReasonCode

pattern CanaryRunStateReasonCode_EXECUTION_FAILURE :: CanaryRunStateReasonCode

CanaryState

newtype CanaryState

Constructors

CanaryState'

Fields

fromCanaryState :: Text

Bundled Patterns

pattern CanaryState_CREATING :: CanaryState

pattern CanaryState_DELETING :: CanaryState

pattern CanaryState_ERROR :: CanaryState

pattern CanaryState_READY :: CanaryState

pattern CanaryState_RUNNING :: CanaryState

pattern CanaryState_STARTING :: CanaryState

pattern CanaryState_STOPPED :: CanaryState

pattern CanaryState_STOPPING :: CanaryState

pattern CanaryState_UPDATING :: CanaryState

CanaryStateReasonCode

newtype CanaryStateReasonCode

Constructors

CanaryStateReasonCode'

Fields

fromCanaryStateReasonCode :: Text

Bundled Patterns

pattern CanaryStateReasonCode_INVALID_PERMISSIONS :: CanaryStateReasonCode

EncryptionMode

newtype EncryptionMode

Constructors

EncryptionMode'

Fields

fromEncryptionMode :: Text

Bundled Patterns

pattern EncryptionMode_SSE_KMS :: EncryptionMode

pattern EncryptionMode_SSE_S3 :: EncryptionMode

ArtifactConfigInput

data ArtifactConfigInput

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

See: newArtifactConfigInput smart constructor.

Constructors

ArtifactConfigInput' (Maybe S3EncryptionConfig)

newArtifactConfigInput :: ArtifactConfigInput

Create a value of ArtifactConfigInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Encryption:ArtifactConfigInput', artifactConfigInput_s3Encryption - A structure that contains the configuration of the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3. Artifact encryption functionality is available only for canaries that use Synthetics runtime version syn-nodejs-puppeteer-3.3 or later. For more information, see Encrypting canary artifacts

ArtifactConfigOutput

data ArtifactConfigOutput

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

See: newArtifactConfigOutput smart constructor.

Constructors

ArtifactConfigOutput' (Maybe S3EncryptionConfig)

newArtifactConfigOutput :: ArtifactConfigOutput

Create a value of ArtifactConfigOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Encryption:ArtifactConfigOutput', artifactConfigOutput_s3Encryption - A structure that contains the configuration of encryption settings for canary artifacts that are stored in Amazon S3.

BaseScreenshot

data BaseScreenshot

A structure representing a screenshot that is used as a baseline during visual monitoring comparisons made by the canary.

See: newBaseScreenshot smart constructor.

Constructors

BaseScreenshot' (Maybe [Text]) Text

newBaseScreenshot

Arguments

:: Text
$sel:screenshotName:BaseScreenshot'

-> BaseScreenshot

Create a value of BaseScreenshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ignoreCoordinates:BaseScreenshot', baseScreenshot_ignoreCoordinates - Coordinates that define the part of a screen to ignore during screenshot comparisons. To obtain the coordinates to use here, use the CloudWatch Logs console to draw the boundaries on the screen. For more information, see {LINK}

$sel:screenshotName:BaseScreenshot', baseScreenshot_screenshotName - The name of the screenshot. This is generated the first time the canary is run after the UpdateCanary operation that specified for this canary to perform visual monitoring.

Canary

data Canary

This structure contains all information about one canary in your account.

See: newCanary smart constructor.

Constructors

Canary' (Maybe CanaryStatus) (Maybe Natural) (Maybe CanaryScheduleOutput) (Maybe Text) (Maybe CanaryRunConfigOutput) (Maybe Text) (Maybe Text) (Maybe Natural) (Maybe ArtifactConfigOutput) (Maybe VpcConfigOutput) (Maybe VisualReferenceOutput) (Maybe Text) (Maybe Text) (Maybe CanaryCodeOutput) (Maybe CanaryTimeline) (Maybe Text) (Maybe (HashMap Text Text))

newCanary :: Canary

Create a value of Canary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Canary', canary_status - A structure that contains information about the canary's status.

$sel:successRetentionPeriodInDays:Canary', canary_successRetentionPeriodInDays - The number of days to retain data about successful runs of this canary.

$sel:schedule:Canary', canary_schedule - A structure that contains information about how often the canary is to run, and when these runs are to stop.

$sel:artifactS3Location:Canary', canary_artifactS3Location - The location in Amazon S3 where Synthetics stores artifacts from the runs of this canary. Artifacts include the log file, screenshots, and HAR files.

$sel:runConfig:Canary', canary_runConfig - Undocumented member.

$sel:executionRoleArn:Canary', canary_executionRoleArn - The ARN of the IAM role used to run the canary. This role must include lambda.amazonaws.com as a principal in the trust policy.

$sel:runtimeVersion:Canary', canary_runtimeVersion - Specifies the runtime version to use for the canary. For more information about runtime versions, see Canary Runtime Versions.

$sel:failureRetentionPeriodInDays:Canary', canary_failureRetentionPeriodInDays - The number of days to retain data about failed runs of this canary.

$sel:artifactConfig:Canary', canary_artifactConfig - A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

$sel:vpcConfig:Canary', canary_vpcConfig - Undocumented member.

$sel:visualReference:Canary', canary_visualReference - If this canary performs visual monitoring by comparing screenshots, this structure contains the ID of the canary run to use as the baseline for screenshots, and the coordinates of any parts of the screen to ignore during the visual monitoring comparison.

$sel:name:Canary', canary_name - The name of the canary.

$sel:id:Canary', canary_id - The unique ID of this canary.

$sel:code:Canary', canary_code - Undocumented member.

$sel:timeline:Canary', canary_timeline - A structure that contains information about when the canary was created, modified, and most recently run.

$sel:engineArn:Canary', canary_engineArn - The ARN of the Lambda function that is used as your canary's engine. For more information about Lambda ARN format, see Resources and Conditions for Lambda Actions.

$sel:tags:Canary', canary_tags - The list of key-value pairs that are associated with the canary.

CanaryCodeInput

data CanaryCodeInput

Use this structure to input your script code for the canary. This structure contains the Lambda handler with the location where the canary should start running the script. If the script is stored in an S3 bucket, the bucket name, key, and version are also included. If the script was passed into the canary directly, the script code is contained in the value of Zipfile.

See: newCanaryCodeInput smart constructor.

Constructors

CanaryCodeInput' (Maybe Text) (Maybe Text) (Maybe Base64) (Maybe Text) Text

newCanaryCodeInput

Arguments

:: Text
$sel:handler:CanaryCodeInput'

-> CanaryCodeInput

Create a value of CanaryCodeInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Key:CanaryCodeInput', canaryCodeInput_s3Key - The S3 key of your script. For more information, see Working with Amazon S3 Objects.

$sel:s3Version:CanaryCodeInput', canaryCodeInput_s3Version - The S3 version ID of your script.

$sel:zipFile:CanaryCodeInput', canaryCodeInput_zipFile - If you input your canary script directly into the canary instead of referring to an S3 location, the value of this parameter is the base64-encoded contents of the .zip file that contains the script. It must be smaller than 256 Kb.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:s3Bucket:CanaryCodeInput', canaryCodeInput_s3Bucket - If your canary script is located in S3, specify the bucket name here. Do not include s3:// as the start of the bucket name.

$sel:handler:CanaryCodeInput', canaryCodeInput_handler - The entry point to use for the source code when running the canary. This value must end with the string .handler. The string is limited to 29 characters or fewer.

CanaryCodeOutput

data CanaryCodeOutput

This structure contains information about the canary's Lambda handler and where its code is stored by CloudWatch Synthetics.

See: newCanaryCodeOutput smart constructor.

Constructors

CanaryCodeOutput' (Maybe Text) (Maybe Text)

newCanaryCodeOutput :: CanaryCodeOutput

Create a value of CanaryCodeOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceLocationArn:CanaryCodeOutput', canaryCodeOutput_sourceLocationArn - The ARN of the Lambda layer where Synthetics stores the canary script code.

$sel:handler:CanaryCodeOutput', canaryCodeOutput_handler - The entry point to use for the source code when running the canary.

CanaryLastRun

data CanaryLastRun

This structure contains information about the most recent run of a single canary.

See: newCanaryLastRun smart constructor.

Constructors

CanaryLastRun' (Maybe Text) (Maybe CanaryRun)

newCanaryLastRun :: CanaryLastRun

Create a value of CanaryLastRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:canaryName:CanaryLastRun', canaryLastRun_canaryName - The name of the canary.

$sel:lastRun:CanaryLastRun', canaryLastRun_lastRun - The results from this canary's most recent run.

CanaryRun

data CanaryRun

This structure contains the details about one run of one canary.

See: newCanaryRun smart constructor.

Constructors

CanaryRun' (Maybe CanaryRunStatus) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe CanaryRunTimeline)

newCanaryRun :: CanaryRun

Create a value of CanaryRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CanaryRun', canaryRun_status - The status of this run.

$sel:artifactS3Location:CanaryRun', canaryRun_artifactS3Location - The location where the canary stored artifacts from the run. Artifacts include the log file, screenshots, and HAR files.

$sel:name:CanaryRun', canaryRun_name - The name of the canary.

$sel:id:CanaryRun', canaryRun_id - A unique ID that identifies this canary run.

$sel:timeline:CanaryRun', canaryRun_timeline - A structure that contains the start and end times of this run.

CanaryRunConfigInput

data CanaryRunConfigInput

A structure that contains input information for a canary run.

See: newCanaryRunConfigInput smart constructor.

Constructors

CanaryRunConfigInput' (Maybe Natural) (Maybe (HashMap Text Text)) (Maybe Bool) (Maybe Natural)

newCanaryRunConfigInput :: CanaryRunConfigInput

Create a value of CanaryRunConfigInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeoutInSeconds:CanaryRunConfigInput', canaryRunConfigInput_timeoutInSeconds - How long the canary is allowed to run before it must stop. You can't set this time to be longer than the frequency of the runs of this canary.

If you omit this field, the frequency of the canary is used as this value, up to a maximum of 14 minutes.

$sel:environmentVariables:CanaryRunConfigInput', canaryRunConfigInput_environmentVariables - Specifies the keys and values to use for any environment variables used in the canary script. Use the following format:

{ "key1" : "value1", "key2" : "value2", ...}

Keys must start with a letter and be at least two characters. The total size of your environment variables cannot exceed 4 KB. You can't specify any Lambda reserved environment variables as the keys for your environment variables. For more information about reserved keys, see Runtime environment variables.

$sel:activeTracing:CanaryRunConfigInput', canaryRunConfigInput_activeTracing - Specifies whether this canary is to use active X-Ray tracing when it runs. Active tracing enables this canary run to be displayed in the ServiceLens and X-Ray service maps even if the canary does not hit an endpoint that has X-Ray tracing enabled. Using X-Ray tracing incurs charges. For more information, see Canaries and X-Ray tracing.

You can enable active tracing only for canaries that use version syn-nodejs-2.0 or later for their canary runtime.

$sel:memoryInMB:CanaryRunConfigInput', canaryRunConfigInput_memoryInMB - The maximum amount of memory available to the canary while it is running, in MB. This value must be a multiple of 64.

CanaryRunConfigOutput

data CanaryRunConfigOutput

A structure that contains information about a canary run.

See: newCanaryRunConfigOutput smart constructor.

Constructors

CanaryRunConfigOutput' (Maybe Natural) (Maybe Bool) (Maybe Natural)

newCanaryRunConfigOutput :: CanaryRunConfigOutput

Create a value of CanaryRunConfigOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeoutInSeconds:CanaryRunConfigOutput', canaryRunConfigOutput_timeoutInSeconds - How long the canary is allowed to run before it must stop.

$sel:activeTracing:CanaryRunConfigOutput', canaryRunConfigOutput_activeTracing - Displays whether this canary run used active X-Ray tracing.

$sel:memoryInMB:CanaryRunConfigOutput', canaryRunConfigOutput_memoryInMB - The maximum amount of memory available to the canary while it is running, in MB. This value must be a multiple of 64.

CanaryRunStatus

data CanaryRunStatus

This structure contains the status information about a canary run.

See: newCanaryRunStatus smart constructor.

Constructors

CanaryRunStatus' (Maybe CanaryRunState) (Maybe Text) (Maybe CanaryRunStateReasonCode)

newCanaryRunStatus :: CanaryRunStatus

Create a value of CanaryRunStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:CanaryRunStatus', canaryRunStatus_state - The current state of the run.

$sel:stateReason:CanaryRunStatus', canaryRunStatus_stateReason - If run of the canary failed, this field contains the reason for the error.

$sel:stateReasonCode:CanaryRunStatus', canaryRunStatus_stateReasonCode - If this value is CANARY_FAILURE, an exception occurred in the canary code. If this value is EXECUTION_FAILURE, an exception occurred in CloudWatch Synthetics.

CanaryRunTimeline

data CanaryRunTimeline

This structure contains the start and end times of a single canary run.

See: newCanaryRunTimeline smart constructor.

Constructors

CanaryRunTimeline' (Maybe POSIX) (Maybe POSIX)

newCanaryRunTimeline :: CanaryRunTimeline

Create a value of CanaryRunTimeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:completed:CanaryRunTimeline', canaryRunTimeline_completed - The end time of the run.

$sel:started:CanaryRunTimeline', canaryRunTimeline_started - The start time of the run.

CanaryScheduleInput

data CanaryScheduleInput

This structure specifies how often a canary is to make runs and the date and time when it should stop making runs.

See: newCanaryScheduleInput smart constructor.

Constructors

CanaryScheduleInput' (Maybe Natural) Text

newCanaryScheduleInput

Arguments

:: Text
$sel:expression:CanaryScheduleInput'

-> CanaryScheduleInput

Create a value of CanaryScheduleInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:durationInSeconds:CanaryScheduleInput', canaryScheduleInput_durationInSeconds - How long, in seconds, for the canary to continue making regular runs according to the schedule in the Expression value. If you specify 0, the canary continues making runs until you stop it. If you omit this field, the default of 0 is used.

$sel:expression:CanaryScheduleInput', canaryScheduleInput_expression - A rate expression or a cron expression that defines how often the canary is to run.

For a rate expression, The syntax is rate(number unit). unit can be minute, minutes, or hour.

For example, rate(1 minute) runs the canary once a minute, rate(10 minutes) runs it once every 10 minutes, and rate(1 hour) runs it once every hour. You can specify a frequency between rate(1 minute) and rate(1 hour).

Specifying rate(0 minute) or rate(0 hour) is a special value that causes the canary to run only once when it is started.

Use cron(expression) to specify a cron expression. You can't schedule a canary to wait for more than a year before running. For information about the syntax for cron expressions, see Scheduling canary runs using cron.

CanaryScheduleOutput

data CanaryScheduleOutput

How long, in seconds, for the canary to continue making regular runs according to the schedule in the Expression value.

See: newCanaryScheduleOutput smart constructor.

Constructors

CanaryScheduleOutput' (Maybe Natural) (Maybe Text)

newCanaryScheduleOutput :: CanaryScheduleOutput

Create a value of CanaryScheduleOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:durationInSeconds:CanaryScheduleOutput', canaryScheduleOutput_durationInSeconds - How long, in seconds, for the canary to continue making regular runs after it was created. The runs are performed according to the schedule in the Expression value.

$sel:expression:CanaryScheduleOutput', canaryScheduleOutput_expression - A rate expression or a cron expression that defines how often the canary is to run.

For a rate expression, The syntax is rate(number unit). unit can be minute, minutes, or hour.

For example, rate(1 minute) runs the canary once a minute, rate(10 minutes) runs it once every 10 minutes, and rate(1 hour) runs it once every hour. You can specify a frequency between rate(1 minute) and rate(1 hour).

Specifying rate(0 minute) or rate(0 hour) is a special value that causes the canary to run only once when it is started.

Use cron(expression) to specify a cron expression. For information about the syntax for cron expressions, see Scheduling canary runs using cron.

CanaryStatus

data CanaryStatus

A structure that contains the current state of the canary.

See: newCanaryStatus smart constructor.

Constructors

CanaryStatus' (Maybe CanaryState) (Maybe Text) (Maybe CanaryStateReasonCode)

newCanaryStatus :: CanaryStatus

Create a value of CanaryStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:CanaryStatus', canaryStatus_state - The current state of the canary.

$sel:stateReason:CanaryStatus', canaryStatus_stateReason - If the canary has insufficient permissions to run, this field provides more details.

$sel:stateReasonCode:CanaryStatus', canaryStatus_stateReasonCode - If the canary cannot run or has failed, this field displays the reason.

CanaryTimeline

data CanaryTimeline

This structure contains information about when the canary was created and modified.

See: newCanaryTimeline smart constructor.

Constructors

CanaryTimeline' (Maybe POSIX) (Maybe POSIX) (Maybe POSIX) (Maybe POSIX)

newCanaryTimeline :: CanaryTimeline

Create a value of CanaryTimeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:created:CanaryTimeline', canaryTimeline_created - The date and time the canary was created.

$sel:lastStarted:CanaryTimeline', canaryTimeline_lastStarted - The date and time that the canary's most recent run started.

$sel:lastStopped:CanaryTimeline', canaryTimeline_lastStopped - The date and time that the canary's most recent run ended.

$sel:lastModified:CanaryTimeline', canaryTimeline_lastModified - The date and time the canary was most recently modified.

RuntimeVersion

data RuntimeVersion

This structure contains information about one canary runtime version. For more information about runtime versions, see Canary Runtime Versions.

See: newRuntimeVersion smart constructor.

Constructors

RuntimeVersion' (Maybe Text) (Maybe POSIX) (Maybe POSIX) (Maybe Text)

newRuntimeVersion :: RuntimeVersion

Create a value of RuntimeVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionName:RuntimeVersion', runtimeVersion_versionName - The name of the runtime version. For a list of valid runtime versions, see Canary Runtime Versions.

$sel:deprecationDate:RuntimeVersion', runtimeVersion_deprecationDate - If this runtime version is deprecated, this value is the date of deprecation.

$sel:releaseDate:RuntimeVersion', runtimeVersion_releaseDate - The date that the runtime version was released.

$sel:description:RuntimeVersion', runtimeVersion_description - A description of the runtime version, created by Amazon.

S3EncryptionConfig

data S3EncryptionConfig

A structure that contains the configuration of encryption-at-rest settings for canary artifacts that the canary uploads to Amazon S3.

For more information, see Encrypting canary artifacts

See: newS3EncryptionConfig smart constructor.

Constructors

S3EncryptionConfig' (Maybe Text) (Maybe EncryptionMode)

newS3EncryptionConfig :: S3EncryptionConfig

Create a value of S3EncryptionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyArn:S3EncryptionConfig', s3EncryptionConfig_kmsKeyArn - The ARN of the customer-managed KMS key to use, if you specify SSE-KMS for EncryptionMode

$sel:encryptionMode:S3EncryptionConfig', s3EncryptionConfig_encryptionMode - The encryption method to use for artifacts created by this canary. Specify SSE_S3 to use server-side encryption (SSE) with an Amazon S3-managed key. Specify SSE-KMS to use server-side encryption with a customer-managed KMS key.

If you omit this parameter, an Amazon Web Services-managed KMS key is used.

VisualReferenceInput

data VisualReferenceInput

An object that specifies what screenshots to use as a baseline for visual monitoring by this canary, and optionally the parts of the screenshots to ignore during the visual monitoring comparison.

Visual monitoring is supported only on canaries running the syn-puppeteer-node-3.2 runtime or later. For more information, see Visual monitoring and Visual monitoring blueprint

See: newVisualReferenceInput smart constructor.

Constructors

VisualReferenceInput' (Maybe [BaseScreenshot]) Text

newVisualReferenceInput

Arguments

:: Text
$sel:baseCanaryRunId:VisualReferenceInput'

-> VisualReferenceInput

Create a value of VisualReferenceInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baseScreenshots:VisualReferenceInput', visualReferenceInput_baseScreenshots - An array of screenshots that will be used as the baseline for visual monitoring in future runs of this canary. If there is a screenshot that you don't want to be used for visual monitoring, remove it from this array.

$sel:baseCanaryRunId:VisualReferenceInput', visualReferenceInput_baseCanaryRunId - Specifies which canary run to use the screenshots from as the baseline for future visual monitoring with this canary. Valid values are nextrun to use the screenshots from the next run after this update is made, lastrun to use the screenshots from the most recent run before this update was made, or the value of Id in the CanaryRun from any past run of this canary.

VisualReferenceOutput

data VisualReferenceOutput

If this canary performs visual monitoring by comparing screenshots, this structure contains the ID of the canary run that is used as the baseline for screenshots, and the coordinates of any parts of those screenshots that are ignored during visual monitoring comparison.

Visual monitoring is supported only on canaries running the syn-puppeteer-node-3.2 runtime or later.

See: newVisualReferenceOutput smart constructor.

Constructors

VisualReferenceOutput' (Maybe [BaseScreenshot]) (Maybe Text)

newVisualReferenceOutput :: VisualReferenceOutput

Create a value of VisualReferenceOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baseScreenshots:VisualReferenceOutput', visualReferenceOutput_baseScreenshots - An array of screenshots that are used as the baseline for comparisons during visual monitoring.

$sel:baseCanaryRunId:VisualReferenceOutput', visualReferenceOutput_baseCanaryRunId - The ID of the canary run that produced the screenshots that are used as the baseline for visual monitoring comparisons during future runs of this canary.

VpcConfigInput

data VpcConfigInput

If this canary is to test an endpoint in a VPC, this structure contains information about the subnets and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.

See: newVpcConfigInput smart constructor.

Constructors

VpcConfigInput' (Maybe [Text]) (Maybe [Text])

newVpcConfigInput :: VpcConfigInput

Create a value of VpcConfigInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:VpcConfigInput', vpcConfigInput_securityGroupIds - The IDs of the security groups for this canary.

$sel:subnetIds:VpcConfigInput', vpcConfigInput_subnetIds - The IDs of the subnets where this canary is to run.

VpcConfigOutput

data VpcConfigOutput

If this canary is to test an endpoint in a VPC, this structure contains information about the subnets and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.

See: newVpcConfigOutput smart constructor.

Constructors

VpcConfigOutput' (Maybe [Text]) (Maybe [Text]) (Maybe Text)

newVpcConfigOutput :: VpcConfigOutput

Create a value of VpcConfigOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:VpcConfigOutput', vpcConfigOutput_securityGroupIds - The IDs of the security groups for this canary.

$sel:subnetIds:VpcConfigOutput', vpcConfigOutput_subnetIds - The IDs of the subnets where this canary is to run.

$sel:vpcId:VpcConfigOutput', vpcConfigOutput_vpcId - The IDs of the VPC where this canary is to run.