Amazonka.Synthetics.DeleteCanary

Description

Permanently deletes the specified canary.

When you delete a canary, resources used and created by the canary are not automatically deleted. After you delete a canary that you do not intend to use again, you should also delete the following:

The Lambda functions and layers used by this canary. These have the prefix cwsyn-MyCanaryName .

The CloudWatch alarms created for this canary. These alarms have a name of Synthetics-SharpDrop-Alarm-MyCanaryName .

Amazon S3 objects and buckets, such as the canary's artifact location.

IAM roles created for the canary. If they were created in the console, these roles have the name role/service-role/CloudWatchSyntheticsRole-MyCanaryName .

CloudWatch Logs log groups created for the canary. These logs groups have the name /aws/lambda/cwsyn-MyCanaryName .

Before you delete a canary, you might want to use GetCanary to display the information about this canary. Make note of the information returned by this operation so that you can delete these resources after you delete the canary.

Creating a Request

data DeleteCanary

See: newDeleteCanary smart constructor.

Constructors

DeleteCanary'

Fields

name :: Text

The name of the canary that you want to delete. To find the names of your canaries, use DescribeCanaries.

newDeleteCanary

Arguments

:: Text
$sel:name:DeleteCanary'

-> DeleteCanary

Create a value of DeleteCanary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteCanary', deleteCanary_name - The name of the canary that you want to delete. To find the names of your canaries, use DescribeCanaries.

Request Lenses

deleteCanary_name :: Lens' DeleteCanary Text

The name of the canary that you want to delete. To find the names of your canaries, use DescribeCanaries.

Destructuring the Response

data DeleteCanaryResponse

See: newDeleteCanaryResponse smart constructor.

Constructors

DeleteCanaryResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteCanaryResponse

Arguments

:: Int
$sel:httpStatus:DeleteCanaryResponse'

-> DeleteCanaryResponse

Create a value of DeleteCanaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteCanaryResponse', deleteCanaryResponse_httpStatus - The response's http status code.

Response Lenses

deleteCanaryResponse_httpStatus :: Lens' DeleteCanaryResponse Int

The response's http status code.