Amazonka.Synthetics.Types

Description

Service Configuration

defaultService :: Service

API version 2017-10-11 of the Amazon Synthetics SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

A parameter could not be validated.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

A conflicting operation is already in progress.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An unknown internal error occurred.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One of the specified resources was not found.

CanaryRunState

newtype CanaryRunState

Constructors

CanaryRunState'

Fields

fromCanaryRunState :: Text

Bundled Patterns

pattern CanaryRunState_FAILED :: CanaryRunState

pattern CanaryRunState_PASSED :: CanaryRunState

pattern CanaryRunState_RUNNING :: CanaryRunState

CanaryRunStateReasonCode

newtype CanaryRunStateReasonCode

Constructors

CanaryRunStateReasonCode'

Fields

fromCanaryRunStateReasonCode :: Text

Bundled Patterns

pattern CanaryRunStateReasonCode_CANARY_FAILURE :: CanaryRunStateReasonCode

pattern CanaryRunStateReasonCode_EXECUTION_FAILURE :: CanaryRunStateReasonCode

CanaryState

newtype CanaryState

Constructors

CanaryState'

Fields

fromCanaryState :: Text

Bundled Patterns

pattern CanaryState_CREATING :: CanaryState

pattern CanaryState_DELETING :: CanaryState

pattern CanaryState_ERROR :: CanaryState

pattern CanaryState_READY :: CanaryState

pattern CanaryState_RUNNING :: CanaryState

pattern CanaryState_STARTING :: CanaryState

pattern CanaryState_STOPPED :: CanaryState

pattern CanaryState_STOPPING :: CanaryState

pattern CanaryState_UPDATING :: CanaryState

CanaryStateReasonCode

newtype CanaryStateReasonCode

Constructors

CanaryStateReasonCode'

Fields

fromCanaryStateReasonCode :: Text

Bundled Patterns

pattern CanaryStateReasonCode_INVALID_PERMISSIONS :: CanaryStateReasonCode

EncryptionMode

newtype EncryptionMode

Constructors

EncryptionMode'

Fields

fromEncryptionMode :: Text

Bundled Patterns

pattern EncryptionMode_SSE_KMS :: EncryptionMode

pattern EncryptionMode_SSE_S3 :: EncryptionMode

ArtifactConfigInput

data ArtifactConfigInput

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

See: newArtifactConfigInput smart constructor.

Constructors

ArtifactConfigInput'

Fields

s3Encryption :: Maybe S3EncryptionConfig

A structure that contains the configuration of the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3. Artifact encryption functionality is available only for canaries that use Synthetics runtime version syn-nodejs-puppeteer-3.3 or later. For more information, see Encrypting canary artifacts

newArtifactConfigInput :: ArtifactConfigInput

Create a value of ArtifactConfigInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Encryption:ArtifactConfigInput', artifactConfigInput_s3Encryption - A structure that contains the configuration of the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3. Artifact encryption functionality is available only for canaries that use Synthetics runtime version syn-nodejs-puppeteer-3.3 or later. For more information, see Encrypting canary artifacts

artifactConfigInput_s3Encryption :: Lens' ArtifactConfigInput (Maybe S3EncryptionConfig)

A structure that contains the configuration of the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3. Artifact encryption functionality is available only for canaries that use Synthetics runtime version syn-nodejs-puppeteer-3.3 or later. For more information, see Encrypting canary artifacts

ArtifactConfigOutput

data ArtifactConfigOutput

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

See: newArtifactConfigOutput smart constructor.

Constructors

ArtifactConfigOutput'

Fields

s3Encryption :: Maybe S3EncryptionConfig

A structure that contains the configuration of encryption settings for canary artifacts that are stored in Amazon S3.

newArtifactConfigOutput :: ArtifactConfigOutput

Create a value of ArtifactConfigOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Encryption:ArtifactConfigOutput', artifactConfigOutput_s3Encryption - A structure that contains the configuration of encryption settings for canary artifacts that are stored in Amazon S3.

artifactConfigOutput_s3Encryption :: Lens' ArtifactConfigOutput (Maybe S3EncryptionConfig)

A structure that contains the configuration of encryption settings for canary artifacts that are stored in Amazon S3.

BaseScreenshot

data BaseScreenshot

A structure representing a screenshot that is used as a baseline during visual monitoring comparisons made by the canary.

See: newBaseScreenshot smart constructor.

Constructors

BaseScreenshot'

Fields

ignoreCoordinates :: Maybe [Text]

Coordinates that define the part of a screen to ignore during screenshot comparisons. To obtain the coordinates to use here, use the CloudWatch Logs console to draw the boundaries on the screen. For more information, see {LINK}

screenshotName :: Text

The name of the screenshot. This is generated the first time the canary is run after the UpdateCanary operation that specified for this canary to perform visual monitoring.

newBaseScreenshot

Arguments

:: Text
$sel:screenshotName:BaseScreenshot'

-> BaseScreenshot

Create a value of BaseScreenshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ignoreCoordinates:BaseScreenshot', baseScreenshot_ignoreCoordinates - Coordinates that define the part of a screen to ignore during screenshot comparisons. To obtain the coordinates to use here, use the CloudWatch Logs console to draw the boundaries on the screen. For more information, see {LINK}

$sel:screenshotName:BaseScreenshot', baseScreenshot_screenshotName - The name of the screenshot. This is generated the first time the canary is run after the UpdateCanary operation that specified for this canary to perform visual monitoring.

baseScreenshot_ignoreCoordinates :: Lens' BaseScreenshot (Maybe [Text])

Coordinates that define the part of a screen to ignore during screenshot comparisons. To obtain the coordinates to use here, use the CloudWatch Logs console to draw the boundaries on the screen. For more information, see {LINK}

baseScreenshot_screenshotName :: Lens' BaseScreenshot Text

The name of the screenshot. This is generated the first time the canary is run after the UpdateCanary operation that specified for this canary to perform visual monitoring.

Canary

data Canary

This structure contains all information about one canary in your account.

See: newCanary smart constructor.

Constructors

Canary'

Fields

status :: Maybe CanaryStatus

A structure that contains information about the canary's status.

successRetentionPeriodInDays :: Maybe Natural

The number of days to retain data about successful runs of this canary.

schedule :: Maybe CanaryScheduleOutput

A structure that contains information about how often the canary is to run, and when these runs are to stop.

artifactS3Location :: Maybe Text

The location in Amazon S3 where Synthetics stores artifacts from the runs of this canary. Artifacts include the log file, screenshots, and HAR files.

runConfig :: Maybe CanaryRunConfigOutput

executionRoleArn :: Maybe Text

The ARN of the IAM role used to run the canary. This role must include lambda.amazonaws.com as a principal in the trust policy.

runtimeVersion :: Maybe Text

Specifies the runtime version to use for the canary. For more information about runtime versions, see Canary Runtime Versions.

failureRetentionPeriodInDays :: Maybe Natural

The number of days to retain data about failed runs of this canary.

artifactConfig :: Maybe ArtifactConfigOutput

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

vpcConfig :: Maybe VpcConfigOutput

visualReference :: Maybe VisualReferenceOutput

If this canary performs visual monitoring by comparing screenshots, this structure contains the ID of the canary run to use as the baseline for screenshots, and the coordinates of any parts of the screen to ignore during the visual monitoring comparison.

name :: Maybe Text

The name of the canary.

id :: Maybe Text

The unique ID of this canary.

code :: Maybe CanaryCodeOutput

timeline :: Maybe CanaryTimeline

A structure that contains information about when the canary was created, modified, and most recently run.

engineArn :: Maybe Text

The ARN of the Lambda function that is used as your canary's engine. For more information about Lambda ARN format, see Resources and Conditions for Lambda Actions.

tags :: Maybe (HashMap Text Text)

The list of key-value pairs that are associated with the canary.

newCanary :: Canary

Create a value of Canary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Canary', canary_status - A structure that contains information about the canary's status.

$sel:successRetentionPeriodInDays:Canary', canary_successRetentionPeriodInDays - The number of days to retain data about successful runs of this canary.

$sel:schedule:Canary', canary_schedule - A structure that contains information about how often the canary is to run, and when these runs are to stop.

$sel:artifactS3Location:Canary', canary_artifactS3Location - The location in Amazon S3 where Synthetics stores artifacts from the runs of this canary. Artifacts include the log file, screenshots, and HAR files.

$sel:runConfig:Canary', canary_runConfig - Undocumented member.

$sel:executionRoleArn:Canary', canary_executionRoleArn - The ARN of the IAM role used to run the canary. This role must include lambda.amazonaws.com as a principal in the trust policy.

$sel:runtimeVersion:Canary', canary_runtimeVersion - Specifies the runtime version to use for the canary. For more information about runtime versions, see Canary Runtime Versions.

$sel:failureRetentionPeriodInDays:Canary', canary_failureRetentionPeriodInDays - The number of days to retain data about failed runs of this canary.

$sel:artifactConfig:Canary', canary_artifactConfig - A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

$sel:vpcConfig:Canary', canary_vpcConfig - Undocumented member.

$sel:visualReference:Canary', canary_visualReference - If this canary performs visual monitoring by comparing screenshots, this structure contains the ID of the canary run to use as the baseline for screenshots, and the coordinates of any parts of the screen to ignore during the visual monitoring comparison.

$sel:name:Canary', canary_name - The name of the canary.

$sel:id:Canary', canary_id - The unique ID of this canary.

$sel:code:Canary', canary_code - Undocumented member.

$sel:timeline:Canary', canary_timeline - A structure that contains information about when the canary was created, modified, and most recently run.

$sel:engineArn:Canary', canary_engineArn - The ARN of the Lambda function that is used as your canary's engine. For more information about Lambda ARN format, see Resources and Conditions for Lambda Actions.

$sel:tags:Canary', canary_tags - The list of key-value pairs that are associated with the canary.

canary_status :: Lens' Canary (Maybe CanaryStatus)

A structure that contains information about the canary's status.

canary_successRetentionPeriodInDays :: Lens' Canary (Maybe Natural)

The number of days to retain data about successful runs of this canary.

canary_schedule :: Lens' Canary (Maybe CanaryScheduleOutput)

A structure that contains information about how often the canary is to run, and when these runs are to stop.

canary_artifactS3Location :: Lens' Canary (Maybe Text)

The location in Amazon S3 where Synthetics stores artifacts from the runs of this canary. Artifacts include the log file, screenshots, and HAR files.

canary_runConfig :: Lens' Canary (Maybe CanaryRunConfigOutput)

Undocumented member.

canary_executionRoleArn :: Lens' Canary (Maybe Text)

The ARN of the IAM role used to run the canary. This role must include lambda.amazonaws.com as a principal in the trust policy.

canary_runtimeVersion :: Lens' Canary (Maybe Text)

Specifies the runtime version to use for the canary. For more information about runtime versions, see Canary Runtime Versions.

canary_failureRetentionPeriodInDays :: Lens' Canary (Maybe Natural)

The number of days to retain data about failed runs of this canary.

canary_artifactConfig :: Lens' Canary (Maybe ArtifactConfigOutput)

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

canary_vpcConfig :: Lens' Canary (Maybe VpcConfigOutput)

Undocumented member.

canary_visualReference :: Lens' Canary (Maybe VisualReferenceOutput)

If this canary performs visual monitoring by comparing screenshots, this structure contains the ID of the canary run to use as the baseline for screenshots, and the coordinates of any parts of the screen to ignore during the visual monitoring comparison.

canary_name :: Lens' Canary (Maybe Text)

The name of the canary.

canary_id :: Lens' Canary (Maybe Text)

The unique ID of this canary.

canary_code :: Lens' Canary (Maybe CanaryCodeOutput)

Undocumented member.

canary_timeline :: Lens' Canary (Maybe CanaryTimeline)

A structure that contains information about when the canary was created, modified, and most recently run.

canary_engineArn :: Lens' Canary (Maybe Text)

The ARN of the Lambda function that is used as your canary's engine. For more information about Lambda ARN format, see Resources and Conditions for Lambda Actions.

canary_tags :: Lens' Canary (Maybe (HashMap Text Text))

The list of key-value pairs that are associated with the canary.

CanaryCodeInput

data CanaryCodeInput

Use this structure to input your script code for the canary. This structure contains the Lambda handler with the location where the canary should start running the script. If the script is stored in an S3 bucket, the bucket name, key, and version are also included. If the script was passed into the canary directly, the script code is contained in the value of Zipfile.

See: newCanaryCodeInput smart constructor.

Constructors

CanaryCodeInput'

Fields

s3Key :: Maybe Text

The S3 key of your script. For more information, see Working with Amazon S3 Objects.

s3Version :: Maybe Text

The S3 version ID of your script.

zipFile :: Maybe Base64

If you input your canary script directly into the canary instead of referring to an S3 location, the value of this parameter is the base64-encoded contents of the .zip file that contains the script. It must be smaller than 256 Kb.

s3Bucket :: Maybe Text

If your canary script is located in S3, specify the bucket name here. Do not include s3:// as the start of the bucket name.

handler :: Text

The entry point to use for the source code when running the canary. This value must end with the string .handler. The string is limited to 29 characters or fewer.

newCanaryCodeInput

Arguments

:: Text
$sel:handler:CanaryCodeInput'

-> CanaryCodeInput

Create a value of CanaryCodeInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Key:CanaryCodeInput', canaryCodeInput_s3Key - The S3 key of your script. For more information, see Working with Amazon S3 Objects.

$sel:s3Version:CanaryCodeInput', canaryCodeInput_s3Version - The S3 version ID of your script.

$sel:zipFile:CanaryCodeInput', canaryCodeInput_zipFile - If you input your canary script directly into the canary instead of referring to an S3 location, the value of this parameter is the base64-encoded contents of the .zip file that contains the script. It must be smaller than 256 Kb.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:s3Bucket:CanaryCodeInput', canaryCodeInput_s3Bucket - If your canary script is located in S3, specify the bucket name here. Do not include s3:// as the start of the bucket name.

$sel:handler:CanaryCodeInput', canaryCodeInput_handler - The entry point to use for the source code when running the canary. This value must end with the string .handler. The string is limited to 29 characters or fewer.

canaryCodeInput_s3Key :: Lens' CanaryCodeInput (Maybe Text)

The S3 key of your script. For more information, see Working with Amazon S3 Objects.

canaryCodeInput_s3Version :: Lens' CanaryCodeInput (Maybe Text)

The S3 version ID of your script.

canaryCodeInput_zipFile :: Lens' CanaryCodeInput (Maybe ByteString)

If you input your canary script directly into the canary instead of referring to an S3 location, the value of this parameter is the base64-encoded contents of the .zip file that contains the script. It must be smaller than 256 Kb.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

canaryCodeInput_s3Bucket :: Lens' CanaryCodeInput (Maybe Text)

If your canary script is located in S3, specify the bucket name here. Do not include s3:// as the start of the bucket name.

canaryCodeInput_handler :: Lens' CanaryCodeInput Text

The entry point to use for the source code when running the canary. This value must end with the string .handler. The string is limited to 29 characters or fewer.

CanaryCodeOutput

data CanaryCodeOutput

This structure contains information about the canary's Lambda handler and where its code is stored by CloudWatch Synthetics.

See: newCanaryCodeOutput smart constructor.

Constructors

CanaryCodeOutput'

Fields

sourceLocationArn :: Maybe Text

The ARN of the Lambda layer where Synthetics stores the canary script code.

handler :: Maybe Text

The entry point to use for the source code when running the canary.

newCanaryCodeOutput :: CanaryCodeOutput

Create a value of CanaryCodeOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceLocationArn:CanaryCodeOutput', canaryCodeOutput_sourceLocationArn - The ARN of the Lambda layer where Synthetics stores the canary script code.

$sel:handler:CanaryCodeOutput', canaryCodeOutput_handler - The entry point to use for the source code when running the canary.

canaryCodeOutput_sourceLocationArn :: Lens' CanaryCodeOutput (Maybe Text)

The ARN of the Lambda layer where Synthetics stores the canary script code.

canaryCodeOutput_handler :: Lens' CanaryCodeOutput (Maybe Text)

The entry point to use for the source code when running the canary.

CanaryLastRun

data CanaryLastRun

This structure contains information about the most recent run of a single canary.

See: newCanaryLastRun smart constructor.

Constructors

CanaryLastRun'

Fields

canaryName :: Maybe Text

The name of the canary.

lastRun :: Maybe CanaryRun

The results from this canary's most recent run.

newCanaryLastRun :: CanaryLastRun

Create a value of CanaryLastRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:canaryName:CanaryLastRun', canaryLastRun_canaryName - The name of the canary.

$sel:lastRun:CanaryLastRun', canaryLastRun_lastRun - The results from this canary's most recent run.

canaryLastRun_canaryName :: Lens' CanaryLastRun (Maybe Text)

The name of the canary.

canaryLastRun_lastRun :: Lens' CanaryLastRun (Maybe CanaryRun)

The results from this canary's most recent run.

CanaryRun

data CanaryRun

This structure contains the details about one run of one canary.

See: newCanaryRun smart constructor.

Constructors

CanaryRun'

Fields

status :: Maybe CanaryRunStatus

The status of this run.

artifactS3Location :: Maybe Text

The location where the canary stored artifacts from the run. Artifacts include the log file, screenshots, and HAR files.

name :: Maybe Text

The name of the canary.

id :: Maybe Text

A unique ID that identifies this canary run.

timeline :: Maybe CanaryRunTimeline

A structure that contains the start and end times of this run.

newCanaryRun :: CanaryRun

Create a value of CanaryRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CanaryRun', canaryRun_status - The status of this run.

$sel:artifactS3Location:CanaryRun', canaryRun_artifactS3Location - The location where the canary stored artifacts from the run. Artifacts include the log file, screenshots, and HAR files.

$sel:name:CanaryRun', canaryRun_name - The name of the canary.

$sel:id:CanaryRun', canaryRun_id - A unique ID that identifies this canary run.

$sel:timeline:CanaryRun', canaryRun_timeline - A structure that contains the start and end times of this run.

canaryRun_status :: Lens' CanaryRun (Maybe CanaryRunStatus)

The status of this run.

canaryRun_artifactS3Location :: Lens' CanaryRun (Maybe Text)

The location where the canary stored artifacts from the run. Artifacts include the log file, screenshots, and HAR files.

canaryRun_name :: Lens' CanaryRun (Maybe Text)

The name of the canary.

canaryRun_id :: Lens' CanaryRun (Maybe Text)

A unique ID that identifies this canary run.

canaryRun_timeline :: Lens' CanaryRun (Maybe CanaryRunTimeline)

A structure that contains the start and end times of this run.

CanaryRunConfigInput

data CanaryRunConfigInput

A structure that contains input information for a canary run.

See: newCanaryRunConfigInput smart constructor.

Constructors

CanaryRunConfigInput'

Fields

timeoutInSeconds :: Maybe Natural

How long the canary is allowed to run before it must stop. You can't set this time to be longer than the frequency of the runs of this canary.

If you omit this field, the frequency of the canary is used as this value, up to a maximum of 14 minutes.

environmentVariables :: Maybe (HashMap Text Text)

Specifies the keys and values to use for any environment variables used in the canary script. Use the following format:

{ "key1" : "value1", "key2" : "value2", ...}

Keys must start with a letter and be at least two characters. The total size of your environment variables cannot exceed 4 KB. You can't specify any Lambda reserved environment variables as the keys for your environment variables. For more information about reserved keys, see Runtime environment variables.

activeTracing :: Maybe Bool

Specifies whether this canary is to use active X-Ray tracing when it runs. Active tracing enables this canary run to be displayed in the ServiceLens and X-Ray service maps even if the canary does not hit an endpoint that has X-Ray tracing enabled. Using X-Ray tracing incurs charges. For more information, see Canaries and X-Ray tracing.

You can enable active tracing only for canaries that use version syn-nodejs-2.0 or later for their canary runtime.

memoryInMB :: Maybe Natural

The maximum amount of memory available to the canary while it is running, in MB. This value must be a multiple of 64.

newCanaryRunConfigInput :: CanaryRunConfigInput

Create a value of CanaryRunConfigInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeoutInSeconds:CanaryRunConfigInput', canaryRunConfigInput_timeoutInSeconds - How long the canary is allowed to run before it must stop. You can't set this time to be longer than the frequency of the runs of this canary.

If you omit this field, the frequency of the canary is used as this value, up to a maximum of 14 minutes.

$sel:environmentVariables:CanaryRunConfigInput', canaryRunConfigInput_environmentVariables - Specifies the keys and values to use for any environment variables used in the canary script. Use the following format:

{ "key1" : "value1", "key2" : "value2", ...}

Keys must start with a letter and be at least two characters. The total size of your environment variables cannot exceed 4 KB. You can't specify any Lambda reserved environment variables as the keys for your environment variables. For more information about reserved keys, see Runtime environment variables.

$sel:activeTracing:CanaryRunConfigInput', canaryRunConfigInput_activeTracing - Specifies whether this canary is to use active X-Ray tracing when it runs. Active tracing enables this canary run to be displayed in the ServiceLens and X-Ray service maps even if the canary does not hit an endpoint that has X-Ray tracing enabled. Using X-Ray tracing incurs charges. For more information, see Canaries and X-Ray tracing.

You can enable active tracing only for canaries that use version syn-nodejs-2.0 or later for their canary runtime.

$sel:memoryInMB:CanaryRunConfigInput', canaryRunConfigInput_memoryInMB - The maximum amount of memory available to the canary while it is running, in MB. This value must be a multiple of 64.

canaryRunConfigInput_timeoutInSeconds :: Lens' CanaryRunConfigInput (Maybe Natural)

How long the canary is allowed to run before it must stop. You can't set this time to be longer than the frequency of the runs of this canary.

If you omit this field, the frequency of the canary is used as this value, up to a maximum of 14 minutes.

canaryRunConfigInput_environmentVariables :: Lens' CanaryRunConfigInput (Maybe (HashMap Text Text))

Specifies the keys and values to use for any environment variables used in the canary script. Use the following format:

{ "key1" : "value1", "key2" : "value2", ...}

Keys must start with a letter and be at least two characters. The total size of your environment variables cannot exceed 4 KB. You can't specify any Lambda reserved environment variables as the keys for your environment variables. For more information about reserved keys, see Runtime environment variables.

canaryRunConfigInput_activeTracing :: Lens' CanaryRunConfigInput (Maybe Bool)

Specifies whether this canary is to use active X-Ray tracing when it runs. Active tracing enables this canary run to be displayed in the ServiceLens and X-Ray service maps even if the canary does not hit an endpoint that has X-Ray tracing enabled. Using X-Ray tracing incurs charges. For more information, see Canaries and X-Ray tracing.

You can enable active tracing only for canaries that use version syn-nodejs-2.0 or later for their canary runtime.

canaryRunConfigInput_memoryInMB :: Lens' CanaryRunConfigInput (Maybe Natural)

The maximum amount of memory available to the canary while it is running, in MB. This value must be a multiple of 64.

CanaryRunConfigOutput

data CanaryRunConfigOutput

A structure that contains information about a canary run.

See: newCanaryRunConfigOutput smart constructor.

Constructors

CanaryRunConfigOutput'

Fields

timeoutInSeconds :: Maybe Natural

How long the canary is allowed to run before it must stop.

activeTracing :: Maybe Bool

Displays whether this canary run used active X-Ray tracing.

memoryInMB :: Maybe Natural

The maximum amount of memory available to the canary while it is running, in MB. This value must be a multiple of 64.

newCanaryRunConfigOutput :: CanaryRunConfigOutput

Create a value of CanaryRunConfigOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeoutInSeconds:CanaryRunConfigOutput', canaryRunConfigOutput_timeoutInSeconds - How long the canary is allowed to run before it must stop.

$sel:activeTracing:CanaryRunConfigOutput', canaryRunConfigOutput_activeTracing - Displays whether this canary run used active X-Ray tracing.

$sel:memoryInMB:CanaryRunConfigOutput', canaryRunConfigOutput_memoryInMB - The maximum amount of memory available to the canary while it is running, in MB. This value must be a multiple of 64.

canaryRunConfigOutput_timeoutInSeconds :: Lens' CanaryRunConfigOutput (Maybe Natural)

How long the canary is allowed to run before it must stop.

canaryRunConfigOutput_activeTracing :: Lens' CanaryRunConfigOutput (Maybe Bool)

Displays whether this canary run used active X-Ray tracing.

canaryRunConfigOutput_memoryInMB :: Lens' CanaryRunConfigOutput (Maybe Natural)

The maximum amount of memory available to the canary while it is running, in MB. This value must be a multiple of 64.

CanaryRunStatus

data CanaryRunStatus

This structure contains the status information about a canary run.

See: newCanaryRunStatus smart constructor.

Constructors

CanaryRunStatus'

Fields

state :: Maybe CanaryRunState

The current state of the run.

stateReason :: Maybe Text

If run of the canary failed, this field contains the reason for the error.

stateReasonCode :: Maybe CanaryRunStateReasonCode

If this value is CANARY_FAILURE, an exception occurred in the canary code. If this value is EXECUTION_FAILURE, an exception occurred in CloudWatch Synthetics.

newCanaryRunStatus :: CanaryRunStatus

Create a value of CanaryRunStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:CanaryRunStatus', canaryRunStatus_state - The current state of the run.

$sel:stateReason:CanaryRunStatus', canaryRunStatus_stateReason - If run of the canary failed, this field contains the reason for the error.

$sel:stateReasonCode:CanaryRunStatus', canaryRunStatus_stateReasonCode - If this value is CANARY_FAILURE, an exception occurred in the canary code. If this value is EXECUTION_FAILURE, an exception occurred in CloudWatch Synthetics.

canaryRunStatus_state :: Lens' CanaryRunStatus (Maybe CanaryRunState)

The current state of the run.

canaryRunStatus_stateReason :: Lens' CanaryRunStatus (Maybe Text)

If run of the canary failed, this field contains the reason for the error.

canaryRunStatus_stateReasonCode :: Lens' CanaryRunStatus (Maybe CanaryRunStateReasonCode)

If this value is CANARY_FAILURE, an exception occurred in the canary code. If this value is EXECUTION_FAILURE, an exception occurred in CloudWatch Synthetics.

CanaryRunTimeline

data CanaryRunTimeline

This structure contains the start and end times of a single canary run.

See: newCanaryRunTimeline smart constructor.

Constructors

CanaryRunTimeline'

Fields

completed :: Maybe POSIX

The end time of the run.

started :: Maybe POSIX

The start time of the run.

newCanaryRunTimeline :: CanaryRunTimeline

Create a value of CanaryRunTimeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:completed:CanaryRunTimeline', canaryRunTimeline_completed - The end time of the run.

$sel:started:CanaryRunTimeline', canaryRunTimeline_started - The start time of the run.

canaryRunTimeline_completed :: Lens' CanaryRunTimeline (Maybe UTCTime)

The end time of the run.

canaryRunTimeline_started :: Lens' CanaryRunTimeline (Maybe UTCTime)

The start time of the run.

CanaryScheduleInput

data CanaryScheduleInput

This structure specifies how often a canary is to make runs and the date and time when it should stop making runs.

See: newCanaryScheduleInput smart constructor.

Constructors

CanaryScheduleInput'

Fields

durationInSeconds :: Maybe Natural

How long, in seconds, for the canary to continue making regular runs according to the schedule in the Expression value. If you specify 0, the canary continues making runs until you stop it. If you omit this field, the default of 0 is used.

expression :: Text

A rate expression or a cron expression that defines how often the canary is to run.

For a rate expression, The syntax is rate(number unit). unit can be minute, minutes, or hour.

For example, rate(1 minute) runs the canary once a minute, rate(10 minutes) runs it once every 10 minutes, and rate(1 hour) runs it once every hour. You can specify a frequency between rate(1 minute) and rate(1 hour).

Specifying rate(0 minute) or rate(0 hour) is a special value that causes the canary to run only once when it is started.

Use cron(expression) to specify a cron expression. You can't schedule a canary to wait for more than a year before running. For information about the syntax for cron expressions, see Scheduling canary runs using cron.

newCanaryScheduleInput

Arguments

:: Text
$sel:expression:CanaryScheduleInput'

-> CanaryScheduleInput

Create a value of CanaryScheduleInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:durationInSeconds:CanaryScheduleInput', canaryScheduleInput_durationInSeconds - How long, in seconds, for the canary to continue making regular runs according to the schedule in the Expression value. If you specify 0, the canary continues making runs until you stop it. If you omit this field, the default of 0 is used.

$sel:expression:CanaryScheduleInput', canaryScheduleInput_expression - A rate expression or a cron expression that defines how often the canary is to run.

For a rate expression, The syntax is rate(number unit). unit can be minute, minutes, or hour.

For example, rate(1 minute) runs the canary once a minute, rate(10 minutes) runs it once every 10 minutes, and rate(1 hour) runs it once every hour. You can specify a frequency between rate(1 minute) and rate(1 hour).

Specifying rate(0 minute) or rate(0 hour) is a special value that causes the canary to run only once when it is started.

Use cron(expression) to specify a cron expression. You can't schedule a canary to wait for more than a year before running. For information about the syntax for cron expressions, see Scheduling canary runs using cron.

canaryScheduleInput_durationInSeconds :: Lens' CanaryScheduleInput (Maybe Natural)

How long, in seconds, for the canary to continue making regular runs according to the schedule in the Expression value. If you specify 0, the canary continues making runs until you stop it. If you omit this field, the default of 0 is used.

canaryScheduleInput_expression :: Lens' CanaryScheduleInput Text

A rate expression or a cron expression that defines how often the canary is to run.

For a rate expression, The syntax is rate(number unit). unit can be minute, minutes, or hour.

For example, rate(1 minute) runs the canary once a minute, rate(10 minutes) runs it once every 10 minutes, and rate(1 hour) runs it once every hour. You can specify a frequency between rate(1 minute) and rate(1 hour).

Specifying rate(0 minute) or rate(0 hour) is a special value that causes the canary to run only once when it is started.

Use cron(expression) to specify a cron expression. You can't schedule a canary to wait for more than a year before running. For information about the syntax for cron expressions, see Scheduling canary runs using cron.

CanaryScheduleOutput

data CanaryScheduleOutput

How long, in seconds, for the canary to continue making regular runs according to the schedule in the Expression value.

See: newCanaryScheduleOutput smart constructor.

Constructors

CanaryScheduleOutput'

Fields

durationInSeconds :: Maybe Natural

How long, in seconds, for the canary to continue making regular runs after it was created. The runs are performed according to the schedule in the Expression value.

expression :: Maybe Text

A rate expression or a cron expression that defines how often the canary is to run.

For a rate expression, The syntax is rate(number unit). unit can be minute, minutes, or hour.

For example, rate(1 minute) runs the canary once a minute, rate(10 minutes) runs it once every 10 minutes, and rate(1 hour) runs it once every hour. You can specify a frequency between rate(1 minute) and rate(1 hour).

Specifying rate(0 minute) or rate(0 hour) is a special value that causes the canary to run only once when it is started.

Use cron(expression) to specify a cron expression. For information about the syntax for cron expressions, see Scheduling canary runs using cron.

newCanaryScheduleOutput :: CanaryScheduleOutput

Create a value of CanaryScheduleOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:durationInSeconds:CanaryScheduleOutput', canaryScheduleOutput_durationInSeconds - How long, in seconds, for the canary to continue making regular runs after it was created. The runs are performed according to the schedule in the Expression value.

$sel:expression:CanaryScheduleOutput', canaryScheduleOutput_expression - A rate expression or a cron expression that defines how often the canary is to run.

For a rate expression, The syntax is rate(number unit). unit can be minute, minutes, or hour.

For example, rate(1 minute) runs the canary once a minute, rate(10 minutes) runs it once every 10 minutes, and rate(1 hour) runs it once every hour. You can specify a frequency between rate(1 minute) and rate(1 hour).

Specifying rate(0 minute) or rate(0 hour) is a special value that causes the canary to run only once when it is started.

Use cron(expression) to specify a cron expression. For information about the syntax for cron expressions, see Scheduling canary runs using cron.

canaryScheduleOutput_durationInSeconds :: Lens' CanaryScheduleOutput (Maybe Natural)

How long, in seconds, for the canary to continue making regular runs after it was created. The runs are performed according to the schedule in the Expression value.

canaryScheduleOutput_expression :: Lens' CanaryScheduleOutput (Maybe Text)

A rate expression or a cron expression that defines how often the canary is to run.

For a rate expression, The syntax is rate(number unit). unit can be minute, minutes, or hour.

For example, rate(1 minute) runs the canary once a minute, rate(10 minutes) runs it once every 10 minutes, and rate(1 hour) runs it once every hour. You can specify a frequency between rate(1 minute) and rate(1 hour).

Specifying rate(0 minute) or rate(0 hour) is a special value that causes the canary to run only once when it is started.

Use cron(expression) to specify a cron expression. For information about the syntax for cron expressions, see Scheduling canary runs using cron.

CanaryStatus

data CanaryStatus

A structure that contains the current state of the canary.

See: newCanaryStatus smart constructor.

Constructors

CanaryStatus'

Fields

state :: Maybe CanaryState

The current state of the canary.

stateReason :: Maybe Text

If the canary has insufficient permissions to run, this field provides more details.

stateReasonCode :: Maybe CanaryStateReasonCode

If the canary cannot run or has failed, this field displays the reason.

newCanaryStatus :: CanaryStatus

Create a value of CanaryStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:CanaryStatus', canaryStatus_state - The current state of the canary.

$sel:stateReason:CanaryStatus', canaryStatus_stateReason - If the canary has insufficient permissions to run, this field provides more details.

$sel:stateReasonCode:CanaryStatus', canaryStatus_stateReasonCode - If the canary cannot run or has failed, this field displays the reason.

canaryStatus_state :: Lens' CanaryStatus (Maybe CanaryState)

The current state of the canary.

canaryStatus_stateReason :: Lens' CanaryStatus (Maybe Text)

If the canary has insufficient permissions to run, this field provides more details.

canaryStatus_stateReasonCode :: Lens' CanaryStatus (Maybe CanaryStateReasonCode)

If the canary cannot run or has failed, this field displays the reason.

CanaryTimeline

data CanaryTimeline

This structure contains information about when the canary was created and modified.

See: newCanaryTimeline smart constructor.

Constructors

CanaryTimeline'

Fields

created :: Maybe POSIX

The date and time the canary was created.

lastStarted :: Maybe POSIX

The date and time that the canary's most recent run started.

lastStopped :: Maybe POSIX

The date and time that the canary's most recent run ended.

lastModified :: Maybe POSIX

The date and time the canary was most recently modified.

newCanaryTimeline :: CanaryTimeline

Create a value of CanaryTimeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:created:CanaryTimeline', canaryTimeline_created - The date and time the canary was created.

$sel:lastStarted:CanaryTimeline', canaryTimeline_lastStarted - The date and time that the canary's most recent run started.

$sel:lastStopped:CanaryTimeline', canaryTimeline_lastStopped - The date and time that the canary's most recent run ended.

$sel:lastModified:CanaryTimeline', canaryTimeline_lastModified - The date and time the canary was most recently modified.

canaryTimeline_created :: Lens' CanaryTimeline (Maybe UTCTime)

The date and time the canary was created.

canaryTimeline_lastStarted :: Lens' CanaryTimeline (Maybe UTCTime)

The date and time that the canary's most recent run started.

canaryTimeline_lastStopped :: Lens' CanaryTimeline (Maybe UTCTime)

The date and time that the canary's most recent run ended.

canaryTimeline_lastModified :: Lens' CanaryTimeline (Maybe UTCTime)

The date and time the canary was most recently modified.

RuntimeVersion

data RuntimeVersion

This structure contains information about one canary runtime version. For more information about runtime versions, see Canary Runtime Versions.

See: newRuntimeVersion smart constructor.

Constructors

RuntimeVersion'

Fields

versionName :: Maybe Text

The name of the runtime version. For a list of valid runtime versions, see Canary Runtime Versions.

deprecationDate :: Maybe POSIX

If this runtime version is deprecated, this value is the date of deprecation.

releaseDate :: Maybe POSIX

The date that the runtime version was released.

description :: Maybe Text

A description of the runtime version, created by Amazon.

newRuntimeVersion :: RuntimeVersion

Create a value of RuntimeVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionName:RuntimeVersion', runtimeVersion_versionName - The name of the runtime version. For a list of valid runtime versions, see Canary Runtime Versions.

$sel:deprecationDate:RuntimeVersion', runtimeVersion_deprecationDate - If this runtime version is deprecated, this value is the date of deprecation.

$sel:releaseDate:RuntimeVersion', runtimeVersion_releaseDate - The date that the runtime version was released.

$sel:description:RuntimeVersion', runtimeVersion_description - A description of the runtime version, created by Amazon.

runtimeVersion_versionName :: Lens' RuntimeVersion (Maybe Text)

The name of the runtime version. For a list of valid runtime versions, see Canary Runtime Versions.

runtimeVersion_deprecationDate :: Lens' RuntimeVersion (Maybe UTCTime)

If this runtime version is deprecated, this value is the date of deprecation.

runtimeVersion_releaseDate :: Lens' RuntimeVersion (Maybe UTCTime)

The date that the runtime version was released.

runtimeVersion_description :: Lens' RuntimeVersion (Maybe Text)

A description of the runtime version, created by Amazon.

S3EncryptionConfig

data S3EncryptionConfig

A structure that contains the configuration of encryption-at-rest settings for canary artifacts that the canary uploads to Amazon S3.

For more information, see Encrypting canary artifacts

See: newS3EncryptionConfig smart constructor.

Constructors

S3EncryptionConfig'

Fields

kmsKeyArn :: Maybe Text

The ARN of the customer-managed KMS key to use, if you specify SSE-KMS for EncryptionMode

encryptionMode :: Maybe EncryptionMode

The encryption method to use for artifacts created by this canary. Specify SSE_S3 to use server-side encryption (SSE) with an Amazon S3-managed key. Specify SSE-KMS to use server-side encryption with a customer-managed KMS key.

If you omit this parameter, an Amazon Web Services-managed KMS key is used.

newS3EncryptionConfig :: S3EncryptionConfig

Create a value of S3EncryptionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyArn:S3EncryptionConfig', s3EncryptionConfig_kmsKeyArn - The ARN of the customer-managed KMS key to use, if you specify SSE-KMS for EncryptionMode

$sel:encryptionMode:S3EncryptionConfig', s3EncryptionConfig_encryptionMode - The encryption method to use for artifacts created by this canary. Specify SSE_S3 to use server-side encryption (SSE) with an Amazon S3-managed key. Specify SSE-KMS to use server-side encryption with a customer-managed KMS key.

If you omit this parameter, an Amazon Web Services-managed KMS key is used.

s3EncryptionConfig_kmsKeyArn :: Lens' S3EncryptionConfig (Maybe Text)

The ARN of the customer-managed KMS key to use, if you specify SSE-KMS for EncryptionMode

s3EncryptionConfig_encryptionMode :: Lens' S3EncryptionConfig (Maybe EncryptionMode)

The encryption method to use for artifacts created by this canary. Specify SSE_S3 to use server-side encryption (SSE) with an Amazon S3-managed key. Specify SSE-KMS to use server-side encryption with a customer-managed KMS key.

If you omit this parameter, an Amazon Web Services-managed KMS key is used.

VisualReferenceInput

data VisualReferenceInput

An object that specifies what screenshots to use as a baseline for visual monitoring by this canary, and optionally the parts of the screenshots to ignore during the visual monitoring comparison.

Visual monitoring is supported only on canaries running the syn-puppeteer-node-3.2 runtime or later. For more information, see Visual monitoring and Visual monitoring blueprint

See: newVisualReferenceInput smart constructor.

Constructors

VisualReferenceInput'

Fields

baseScreenshots :: Maybe [BaseScreenshot]

An array of screenshots that will be used as the baseline for visual monitoring in future runs of this canary. If there is a screenshot that you don't want to be used for visual monitoring, remove it from this array.

baseCanaryRunId :: Text

Specifies which canary run to use the screenshots from as the baseline for future visual monitoring with this canary. Valid values are nextrun to use the screenshots from the next run after this update is made, lastrun to use the screenshots from the most recent run before this update was made, or the value of Id in the CanaryRun from any past run of this canary.

newVisualReferenceInput

Arguments

:: Text
$sel:baseCanaryRunId:VisualReferenceInput'

-> VisualReferenceInput

Create a value of VisualReferenceInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baseScreenshots:VisualReferenceInput', visualReferenceInput_baseScreenshots - An array of screenshots that will be used as the baseline for visual monitoring in future runs of this canary. If there is a screenshot that you don't want to be used for visual monitoring, remove it from this array.

$sel:baseCanaryRunId:VisualReferenceInput', visualReferenceInput_baseCanaryRunId - Specifies which canary run to use the screenshots from as the baseline for future visual monitoring with this canary. Valid values are nextrun to use the screenshots from the next run after this update is made, lastrun to use the screenshots from the most recent run before this update was made, or the value of Id in the CanaryRun from any past run of this canary.

visualReferenceInput_baseScreenshots :: Lens' VisualReferenceInput (Maybe [BaseScreenshot])

An array of screenshots that will be used as the baseline for visual monitoring in future runs of this canary. If there is a screenshot that you don't want to be used for visual monitoring, remove it from this array.

visualReferenceInput_baseCanaryRunId :: Lens' VisualReferenceInput Text

Specifies which canary run to use the screenshots from as the baseline for future visual monitoring with this canary. Valid values are nextrun to use the screenshots from the next run after this update is made, lastrun to use the screenshots from the most recent run before this update was made, or the value of Id in the CanaryRun from any past run of this canary.

VisualReferenceOutput

data VisualReferenceOutput

If this canary performs visual monitoring by comparing screenshots, this structure contains the ID of the canary run that is used as the baseline for screenshots, and the coordinates of any parts of those screenshots that are ignored during visual monitoring comparison.

Visual monitoring is supported only on canaries running the syn-puppeteer-node-3.2 runtime or later.

See: newVisualReferenceOutput smart constructor.

Constructors

VisualReferenceOutput'

Fields

baseScreenshots :: Maybe [BaseScreenshot]

An array of screenshots that are used as the baseline for comparisons during visual monitoring.

baseCanaryRunId :: Maybe Text

The ID of the canary run that produced the screenshots that are used as the baseline for visual monitoring comparisons during future runs of this canary.

newVisualReferenceOutput :: VisualReferenceOutput

Create a value of VisualReferenceOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baseScreenshots:VisualReferenceOutput', visualReferenceOutput_baseScreenshots - An array of screenshots that are used as the baseline for comparisons during visual monitoring.

$sel:baseCanaryRunId:VisualReferenceOutput', visualReferenceOutput_baseCanaryRunId - The ID of the canary run that produced the screenshots that are used as the baseline for visual monitoring comparisons during future runs of this canary.

visualReferenceOutput_baseScreenshots :: Lens' VisualReferenceOutput (Maybe [BaseScreenshot])

An array of screenshots that are used as the baseline for comparisons during visual monitoring.

visualReferenceOutput_baseCanaryRunId :: Lens' VisualReferenceOutput (Maybe Text)

The ID of the canary run that produced the screenshots that are used as the baseline for visual monitoring comparisons during future runs of this canary.

VpcConfigInput

data VpcConfigInput

If this canary is to test an endpoint in a VPC, this structure contains information about the subnets and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.

See: newVpcConfigInput smart constructor.

Constructors

VpcConfigInput'

Fields

securityGroupIds :: Maybe [Text]

The IDs of the security groups for this canary.

subnetIds :: Maybe [Text]

The IDs of the subnets where this canary is to run.

newVpcConfigInput :: VpcConfigInput

Create a value of VpcConfigInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:VpcConfigInput', vpcConfigInput_securityGroupIds - The IDs of the security groups for this canary.

$sel:subnetIds:VpcConfigInput', vpcConfigInput_subnetIds - The IDs of the subnets where this canary is to run.

vpcConfigInput_securityGroupIds :: Lens' VpcConfigInput (Maybe [Text])

The IDs of the security groups for this canary.

vpcConfigInput_subnetIds :: Lens' VpcConfigInput (Maybe [Text])

The IDs of the subnets where this canary is to run.

VpcConfigOutput

data VpcConfigOutput

If this canary is to test an endpoint in a VPC, this structure contains information about the subnets and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.

See: newVpcConfigOutput smart constructor.

Constructors

VpcConfigOutput'

Fields

securityGroupIds :: Maybe [Text]

The IDs of the security groups for this canary.

subnetIds :: Maybe [Text]

The IDs of the subnets where this canary is to run.

vpcId :: Maybe Text

The IDs of the VPC where this canary is to run.

newVpcConfigOutput :: VpcConfigOutput

Create a value of VpcConfigOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:VpcConfigOutput', vpcConfigOutput_securityGroupIds - The IDs of the security groups for this canary.

$sel:subnetIds:VpcConfigOutput', vpcConfigOutput_subnetIds - The IDs of the subnets where this canary is to run.

$sel:vpcId:VpcConfigOutput', vpcConfigOutput_vpcId - The IDs of the VPC where this canary is to run.

vpcConfigOutput_securityGroupIds :: Lens' VpcConfigOutput (Maybe [Text])

The IDs of the security groups for this canary.

vpcConfigOutput_subnetIds :: Lens' VpcConfigOutput (Maybe [Text])

The IDs of the subnets where this canary is to run.

vpcConfigOutput_vpcId :: Lens' VpcConfigOutput (Maybe Text)

The IDs of the VPC where this canary is to run.