Amazonka.Synthetics.Types.Canary

Description

Documentation

data Canary

This structure contains all information about one canary in your account.

See: newCanary smart constructor.

Constructors

Canary'

Fields

status :: Maybe CanaryStatus

A structure that contains information about the canary's status.

successRetentionPeriodInDays :: Maybe Natural

The number of days to retain data about successful runs of this canary.

schedule :: Maybe CanaryScheduleOutput

A structure that contains information about how often the canary is to run, and when these runs are to stop.

artifactS3Location :: Maybe Text

The location in Amazon S3 where Synthetics stores artifacts from the runs of this canary. Artifacts include the log file, screenshots, and HAR files.

runConfig :: Maybe CanaryRunConfigOutput

executionRoleArn :: Maybe Text

The ARN of the IAM role used to run the canary. This role must include lambda.amazonaws.com as a principal in the trust policy.

runtimeVersion :: Maybe Text

Specifies the runtime version to use for the canary. For more information about runtime versions, see Canary Runtime Versions.

failureRetentionPeriodInDays :: Maybe Natural

The number of days to retain data about failed runs of this canary.

artifactConfig :: Maybe ArtifactConfigOutput

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

vpcConfig :: Maybe VpcConfigOutput

visualReference :: Maybe VisualReferenceOutput

If this canary performs visual monitoring by comparing screenshots, this structure contains the ID of the canary run to use as the baseline for screenshots, and the coordinates of any parts of the screen to ignore during the visual monitoring comparison.

name :: Maybe Text

The name of the canary.

id :: Maybe Text

The unique ID of this canary.

code :: Maybe CanaryCodeOutput

timeline :: Maybe CanaryTimeline

A structure that contains information about when the canary was created, modified, and most recently run.

engineArn :: Maybe Text

The ARN of the Lambda function that is used as your canary's engine. For more information about Lambda ARN format, see Resources and Conditions for Lambda Actions.

tags :: Maybe (HashMap Text Text)

The list of key-value pairs that are associated with the canary.

newCanary :: Canary

Create a value of Canary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Canary', canary_status - A structure that contains information about the canary's status.

$sel:successRetentionPeriodInDays:Canary', canary_successRetentionPeriodInDays - The number of days to retain data about successful runs of this canary.

$sel:schedule:Canary', canary_schedule - A structure that contains information about how often the canary is to run, and when these runs are to stop.

$sel:artifactS3Location:Canary', canary_artifactS3Location - The location in Amazon S3 where Synthetics stores artifacts from the runs of this canary. Artifacts include the log file, screenshots, and HAR files.

$sel:runConfig:Canary', canary_runConfig - Undocumented member.

$sel:executionRoleArn:Canary', canary_executionRoleArn - The ARN of the IAM role used to run the canary. This role must include lambda.amazonaws.com as a principal in the trust policy.

$sel:runtimeVersion:Canary', canary_runtimeVersion - Specifies the runtime version to use for the canary. For more information about runtime versions, see Canary Runtime Versions.

$sel:failureRetentionPeriodInDays:Canary', canary_failureRetentionPeriodInDays - The number of days to retain data about failed runs of this canary.

$sel:artifactConfig:Canary', canary_artifactConfig - A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

$sel:vpcConfig:Canary', canary_vpcConfig - Undocumented member.

$sel:visualReference:Canary', canary_visualReference - If this canary performs visual monitoring by comparing screenshots, this structure contains the ID of the canary run to use as the baseline for screenshots, and the coordinates of any parts of the screen to ignore during the visual monitoring comparison.

$sel:name:Canary', canary_name - The name of the canary.

$sel:id:Canary', canary_id - The unique ID of this canary.

$sel:code:Canary', canary_code - Undocumented member.

$sel:timeline:Canary', canary_timeline - A structure that contains information about when the canary was created, modified, and most recently run.

$sel:engineArn:Canary', canary_engineArn - The ARN of the Lambda function that is used as your canary's engine. For more information about Lambda ARN format, see Resources and Conditions for Lambda Actions.

$sel:tags:Canary', canary_tags - The list of key-value pairs that are associated with the canary.

canary_status :: Lens' Canary (Maybe CanaryStatus)

A structure that contains information about the canary's status.

canary_successRetentionPeriodInDays :: Lens' Canary (Maybe Natural)

The number of days to retain data about successful runs of this canary.

canary_schedule :: Lens' Canary (Maybe CanaryScheduleOutput)

A structure that contains information about how often the canary is to run, and when these runs are to stop.

canary_artifactS3Location :: Lens' Canary (Maybe Text)

The location in Amazon S3 where Synthetics stores artifacts from the runs of this canary. Artifacts include the log file, screenshots, and HAR files.

canary_runConfig :: Lens' Canary (Maybe CanaryRunConfigOutput)

Undocumented member.

canary_executionRoleArn :: Lens' Canary (Maybe Text)

The ARN of the IAM role used to run the canary. This role must include lambda.amazonaws.com as a principal in the trust policy.

canary_runtimeVersion :: Lens' Canary (Maybe Text)

Specifies the runtime version to use for the canary. For more information about runtime versions, see Canary Runtime Versions.

canary_failureRetentionPeriodInDays :: Lens' Canary (Maybe Natural)

The number of days to retain data about failed runs of this canary.

canary_artifactConfig :: Lens' Canary (Maybe ArtifactConfigOutput)

A structure that contains the configuration for canary artifacts, including the encryption-at-rest settings for artifacts that the canary uploads to Amazon S3.

canary_vpcConfig :: Lens' Canary (Maybe VpcConfigOutput)

Undocumented member.

canary_visualReference :: Lens' Canary (Maybe VisualReferenceOutput)

If this canary performs visual monitoring by comparing screenshots, this structure contains the ID of the canary run to use as the baseline for screenshots, and the coordinates of any parts of the screen to ignore during the visual monitoring comparison.

canary_name :: Lens' Canary (Maybe Text)

The name of the canary.

canary_id :: Lens' Canary (Maybe Text)

The unique ID of this canary.

canary_code :: Lens' Canary (Maybe CanaryCodeOutput)

Undocumented member.

canary_timeline :: Lens' Canary (Maybe CanaryTimeline)

A structure that contains information about when the canary was created, modified, and most recently run.

canary_engineArn :: Lens' Canary (Maybe Text)

The ARN of the Lambda function that is used as your canary's engine. For more information about Lambda ARN format, see Resources and Conditions for Lambda Actions.

canary_tags :: Lens' Canary (Maybe (HashMap Text Text))

The list of key-value pairs that are associated with the canary.