Amazonka.Textract.Types.Block

Description

Documentation

data Block

A Block represents items that are recognized in a document within a group of pixels close to each other. The information returned in a Block object depends on the type of operation. In text detection for documents (for example DetectDocumentText), you get information about the detected words and lines of text. In text analysis (for example AnalyzeDocument), you can also get information about the fields, tables, and selection elements that are detected in the document.

An array of Block objects is returned by both synchronous and asynchronous operations. In synchronous operations, such as DetectDocumentText, the array of Block objects is the entire set of results. In asynchronous operations, such as GetDocumentAnalysis, the array is returned over one or more responses.

For more information, see How Amazon Textract Works.

See: newBlock smart constructor.

Constructors

Block'

Fields

columnSpan :: Maybe Natural

The number of columns that a table cell spans. Currently this value is always 1, even if the number of columns spanned is greater than 1. ColumnSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

text :: Maybe Text

The word or line of text that's recognized by Amazon Textract.

entityTypes :: Maybe [EntityType]

The type of entity. The following can be returned:

KEY - An identifier for a field on the document.

VALUE - The field text.

EntityTypes isn't returned by DetectDocumentText and GetDocumentTextDetection.

columnIndex :: Maybe Natural

The column in which a table cell appears. The first column position is 1. ColumnIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

page :: Maybe Natural

The page on which a block was detected. Page is returned by asynchronous operations. Page values greater than 1 are only returned for multipage documents that are in PDF format. A scanned image (JPEG/PNG), even if it contains multiple document pages, is considered to be a single-page document. The value of Page is always 1. Synchronous operations don't return Page because every input document is considered to be a single-page document.

rowSpan :: Maybe Natural

The number of rows that a table cell spans. Currently this value is always 1, even if the number of rows spanned is greater than 1. RowSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

selectionStatus :: Maybe SelectionStatus

The selection status of a selection element, such as an option button or check box.

rowIndex :: Maybe Natural

The row in which a table cell is located. The first row position is 1. RowIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

confidence :: Maybe Double

The confidence score that Amazon Textract has in the accuracy of the recognized text and the accuracy of the geometry points around the recognized text.

relationships :: Maybe [Relationship]

A list of child blocks of the current block. For example, a LINE object has child blocks for each WORD block that's part of the line of text. There aren't Relationship objects in the list for relationships that don't exist, such as when the current block has no child blocks. The list size can be the following:

0 - The block has no child blocks.

1 - The block has child blocks.

geometry :: Maybe Geometry

The location of the recognized text on the image. It includes an axis-aligned, coarse bounding box that surrounds the text, and a finer-grain polygon for more accurate spatial information.

textType :: Maybe TextType

The kind of text that Amazon Textract has detected. Can check for handwritten text and printed text.

id :: Maybe Text

The identifier for the recognized text. The identifier is only unique for a single operation.

blockType :: Maybe BlockType

The type of text item that's recognized. In operations for text detection, the following types are returned:

PAGE - Contains a list of the LINE Block objects that are detected on a document page.

WORD - A word detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

In text analysis operations, the following types are returned:

PAGE - Contains a list of child Block objects that are detected on a document page.

KEY_VALUE_SET - Stores the KEY and VALUE Block objects for linked text that's detected on a document page. Use the EntityType field to determine if a KEY_VALUE_SET object is a KEY Block object or a VALUE Block object.

WORD - A word that's detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

TABLE - A table that's detected on a document page. A table is grid-based information with two or more rows or columns, with a cell span of one row and one column each.

CELL - A cell within a detected table. The cell is the parent of the block that contains the text in the cell.

SELECTION_ELEMENT - A selection element such as an option button (radio button) or a check box that's detected on a document page. Use the value of SelectionStatus to determine the status of the selection element.

newBlock :: Block

Create a value of Block with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:columnSpan:Block', block_columnSpan - The number of columns that a table cell spans. Currently this value is always 1, even if the number of columns spanned is greater than 1. ColumnSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

$sel:text:Block', block_text - The word or line of text that's recognized by Amazon Textract.

$sel:entityTypes:Block', block_entityTypes - The type of entity. The following can be returned:

KEY - An identifier for a field on the document.

VALUE - The field text.

EntityTypes isn't returned by DetectDocumentText and GetDocumentTextDetection.

$sel:columnIndex:Block', block_columnIndex - The column in which a table cell appears. The first column position is 1. ColumnIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

$sel:page:Block', block_page - The page on which a block was detected. Page is returned by asynchronous operations. Page values greater than 1 are only returned for multipage documents that are in PDF format. A scanned image (JPEG/PNG), even if it contains multiple document pages, is considered to be a single-page document. The value of Page is always 1. Synchronous operations don't return Page because every input document is considered to be a single-page document.

$sel:rowSpan:Block', block_rowSpan - The number of rows that a table cell spans. Currently this value is always 1, even if the number of rows spanned is greater than 1. RowSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

$sel:selectionStatus:Block', block_selectionStatus - The selection status of a selection element, such as an option button or check box.

$sel:rowIndex:Block', block_rowIndex - The row in which a table cell is located. The first row position is 1. RowIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

$sel:confidence:Block', block_confidence - The confidence score that Amazon Textract has in the accuracy of the recognized text and the accuracy of the geometry points around the recognized text.

$sel:relationships:Block', block_relationships - A list of child blocks of the current block. For example, a LINE object has child blocks for each WORD block that's part of the line of text. There aren't Relationship objects in the list for relationships that don't exist, such as when the current block has no child blocks. The list size can be the following:

0 - The block has no child blocks.

1 - The block has child blocks.

$sel:geometry:Block', block_geometry - The location of the recognized text on the image. It includes an axis-aligned, coarse bounding box that surrounds the text, and a finer-grain polygon for more accurate spatial information.

$sel:textType:Block', block_textType - The kind of text that Amazon Textract has detected. Can check for handwritten text and printed text.

$sel:id:Block', block_id - The identifier for the recognized text. The identifier is only unique for a single operation.

$sel:blockType:Block', block_blockType - The type of text item that's recognized. In operations for text detection, the following types are returned:

PAGE - Contains a list of the LINE Block objects that are detected on a document page.

WORD - A word detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

In text analysis operations, the following types are returned:

PAGE - Contains a list of child Block objects that are detected on a document page.

KEY_VALUE_SET - Stores the KEY and VALUE Block objects for linked text that's detected on a document page. Use the EntityType field to determine if a KEY_VALUE_SET object is a KEY Block object or a VALUE Block object.

WORD - A word that's detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

TABLE - A table that's detected on a document page. A table is grid-based information with two or more rows or columns, with a cell span of one row and one column each.

CELL - A cell within a detected table. The cell is the parent of the block that contains the text in the cell.

SELECTION_ELEMENT - A selection element such as an option button (radio button) or a check box that's detected on a document page. Use the value of SelectionStatus to determine the status of the selection element.

block_columnSpan :: Lens' Block (Maybe Natural)

The number of columns that a table cell spans. Currently this value is always 1, even if the number of columns spanned is greater than 1. ColumnSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_text :: Lens' Block (Maybe Text)

The word or line of text that's recognized by Amazon Textract.

block_entityTypes :: Lens' Block (Maybe [EntityType])

The type of entity. The following can be returned:

KEY - An identifier for a field on the document.

VALUE - The field text.

EntityTypes isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_columnIndex :: Lens' Block (Maybe Natural)

The column in which a table cell appears. The first column position is 1. ColumnIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_page :: Lens' Block (Maybe Natural)

The page on which a block was detected. Page is returned by asynchronous operations. Page values greater than 1 are only returned for multipage documents that are in PDF format. A scanned image (JPEG/PNG), even if it contains multiple document pages, is considered to be a single-page document. The value of Page is always 1. Synchronous operations don't return Page because every input document is considered to be a single-page document.

block_rowSpan :: Lens' Block (Maybe Natural)

The number of rows that a table cell spans. Currently this value is always 1, even if the number of rows spanned is greater than 1. RowSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_selectionStatus :: Lens' Block (Maybe SelectionStatus)

The selection status of a selection element, such as an option button or check box.

block_rowIndex :: Lens' Block (Maybe Natural)

The row in which a table cell is located. The first row position is 1. RowIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

block_confidence :: Lens' Block (Maybe Double)

The confidence score that Amazon Textract has in the accuracy of the recognized text and the accuracy of the geometry points around the recognized text.

block_relationships :: Lens' Block (Maybe [Relationship])

A list of child blocks of the current block. For example, a LINE object has child blocks for each WORD block that's part of the line of text. There aren't Relationship objects in the list for relationships that don't exist, such as when the current block has no child blocks. The list size can be the following:

0 - The block has no child blocks.

1 - The block has child blocks.

block_geometry :: Lens' Block (Maybe Geometry)

The location of the recognized text on the image. It includes an axis-aligned, coarse bounding box that surrounds the text, and a finer-grain polygon for more accurate spatial information.

block_textType :: Lens' Block (Maybe TextType)

The kind of text that Amazon Textract has detected. Can check for handwritten text and printed text.

block_id :: Lens' Block (Maybe Text)

The identifier for the recognized text. The identifier is only unique for a single operation.

block_blockType :: Lens' Block (Maybe BlockType)

The type of text item that's recognized. In operations for text detection, the following types are returned:

PAGE - Contains a list of the LINE Block objects that are detected on a document page.

WORD - A word detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

In text analysis operations, the following types are returned:

PAGE - Contains a list of child Block objects that are detected on a document page.

KEY_VALUE_SET - Stores the KEY and VALUE Block objects for linked text that's detected on a document page. Use the EntityType field to determine if a KEY_VALUE_SET object is a KEY Block object or a VALUE Block object.

WORD - A word that's detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

TABLE - A table that's detected on a document page. A table is grid-based information with two or more rows or columns, with a cell span of one row and one column each.

CELL - A cell within a detected table. The cell is the parent of the block that contains the text in the cell.

SELECTION_ELEMENT - A selection element such as an option button (radio button) or a check box that's detected on a document page. Use the value of SelectionStatus to determine the status of the selection element.