Amazonka.Textract.Types.BoundingBox

Description

Documentation

data BoundingBox

The bounding box around the detected page, text, key-value pair, table, table cell, or selection element on a document page. The left (x-coordinate) and top (y-coordinate) are coordinates that represent the top and left sides of the bounding box. Note that the upper-left corner of the image is the origin (0,0).

The top and left values returned are ratios of the overall document page size. For example, if the input image is 700 x 200 pixels, and the top-left coordinate of the bounding box is 350 x 50 pixels, the API returns a left value of 0.5 (350/700) and a top value of 0.25 (50/200).

The width and height values represent the dimensions of the bounding box as a ratio of the overall document page dimension. For example, if the document page size is 700 x 200 pixels, and the bounding box width is 70 pixels, the width returned is 0.1.

See: newBoundingBox smart constructor.

Constructors

BoundingBox'

Fields

height :: Maybe Double

The height of the bounding box as a ratio of the overall document page height.

left :: Maybe Double

The left coordinate of the bounding box as a ratio of overall document page width.

width :: Maybe Double

The width of the bounding box as a ratio of the overall document page width.

top :: Maybe Double

The top coordinate of the bounding box as a ratio of overall document page height.

newBoundingBox :: BoundingBox

Create a value of BoundingBox with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:height:BoundingBox', boundingBox_height - The height of the bounding box as a ratio of the overall document page height.

$sel:left:BoundingBox', boundingBox_left - The left coordinate of the bounding box as a ratio of overall document page width.

$sel:width:BoundingBox', boundingBox_width - The width of the bounding box as a ratio of the overall document page width.

$sel:top:BoundingBox', boundingBox_top - The top coordinate of the bounding box as a ratio of overall document page height.

boundingBox_height :: Lens' BoundingBox (Maybe Double)

The height of the bounding box as a ratio of the overall document page height.

boundingBox_left :: Lens' BoundingBox (Maybe Double)

The left coordinate of the bounding box as a ratio of overall document page width.

boundingBox_width :: Lens' BoundingBox (Maybe Double)

The width of the bounding box as a ratio of the overall document page width.

boundingBox_top :: Lens' BoundingBox (Maybe Double)

The top coordinate of the bounding box as a ratio of overall document page height.