Amazonka.Textract.Types.ExpenseDetection

Description

Documentation

data ExpenseDetection

An object used to store information about the Value or Label detected by Amazon Textract.

See: newExpenseDetection smart constructor.

Constructors

ExpenseDetection'

Fields

text :: Maybe Text

The word or line of text recognized by Amazon Textract

confidence :: Maybe Double

The confidence in detection, as a percentage

geometry :: Maybe Geometry

newExpenseDetection :: ExpenseDetection

Create a value of ExpenseDetection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:text:ExpenseDetection', expenseDetection_text - The word or line of text recognized by Amazon Textract

$sel:confidence:ExpenseDetection', expenseDetection_confidence - The confidence in detection, as a percentage

$sel:geometry:ExpenseDetection', expenseDetection_geometry - Undocumented member.

expenseDetection_text :: Lens' ExpenseDetection (Maybe Text)

The word or line of text recognized by Amazon Textract

expenseDetection_confidence :: Lens' ExpenseDetection (Maybe Double)

The confidence in detection, as a percentage

expenseDetection_geometry :: Lens' ExpenseDetection (Maybe Geometry)

Undocumented member.