Amazonka.Textract.Types.HumanLoopConfig

Description

Documentation

data HumanLoopConfig

Sets up the human review workflow the document will be sent to if one of the conditions is met. You can also set certain attributes of the image before review.

See: newHumanLoopConfig smart constructor.

Constructors

HumanLoopConfig'

Fields

dataAttributes :: Maybe HumanLoopDataAttributes

Sets attributes of the input data.

humanLoopName :: Text

The name of the human workflow used for this image. This should be kept unique within a region.

flowDefinitionArn :: Text

The Amazon Resource Name (ARN) of the flow definition.

newHumanLoopConfig

Arguments

:: Text
$sel:humanLoopName:HumanLoopConfig'

-> Text
$sel:flowDefinitionArn:HumanLoopConfig'

-> HumanLoopConfig

Create a value of HumanLoopConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataAttributes:HumanLoopConfig', humanLoopConfig_dataAttributes - Sets attributes of the input data.

$sel:humanLoopName:HumanLoopConfig', humanLoopConfig_humanLoopName - The name of the human workflow used for this image. This should be kept unique within a region.

$sel:flowDefinitionArn:HumanLoopConfig', humanLoopConfig_flowDefinitionArn - The Amazon Resource Name (ARN) of the flow definition.

humanLoopConfig_dataAttributes :: Lens' HumanLoopConfig (Maybe HumanLoopDataAttributes)

Sets attributes of the input data.

humanLoopConfig_humanLoopName :: Lens' HumanLoopConfig Text

The name of the human workflow used for this image. This should be kept unique within a region.

humanLoopConfig_flowDefinitionArn :: Lens' HumanLoopConfig Text

The Amazon Resource Name (ARN) of the flow definition.