Amazonka.Transcribe.DeleteMedicalTranscriptionJob

Description

Deletes a transcription job generated by Amazon Transcribe Medical and any related information.

Creating a Request

data DeleteMedicalTranscriptionJob

See: newDeleteMedicalTranscriptionJob smart constructor.

Constructors

DeleteMedicalTranscriptionJob'

Fields

medicalTranscriptionJobName :: Text

The name you provide to the DeleteMedicalTranscriptionJob object to delete a transcription job.

newDeleteMedicalTranscriptionJob

Arguments

:: Text
$sel:medicalTranscriptionJobName:DeleteMedicalTranscriptionJob'

-> DeleteMedicalTranscriptionJob

Create a value of DeleteMedicalTranscriptionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:medicalTranscriptionJobName:DeleteMedicalTranscriptionJob', deleteMedicalTranscriptionJob_medicalTranscriptionJobName - The name you provide to the DeleteMedicalTranscriptionJob object to delete a transcription job.

Request Lenses

deleteMedicalTranscriptionJob_medicalTranscriptionJobName :: Lens' DeleteMedicalTranscriptionJob Text

The name you provide to the DeleteMedicalTranscriptionJob object to delete a transcription job.

Destructuring the Response

data DeleteMedicalTranscriptionJobResponse

See: newDeleteMedicalTranscriptionJobResponse smart constructor.

Constructors

DeleteMedicalTranscriptionJobResponse'

newDeleteMedicalTranscriptionJobResponse :: DeleteMedicalTranscriptionJobResponse

Create a value of DeleteMedicalTranscriptionJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.