Amazonka.Transcribe.GetVocabulary

Description

Gets information about a vocabulary.

Creating a Request

data GetVocabulary

See: newGetVocabulary smart constructor.

Constructors

GetVocabulary'

Fields

vocabularyName :: Text

The name of the vocabulary to return information about. The name is case sensitive.

newGetVocabulary

Arguments

:: Text
$sel:vocabularyName:GetVocabulary'

-> GetVocabulary

Create a value of GetVocabulary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vocabularyName:GetVocabulary', getVocabulary_vocabularyName - The name of the vocabulary to return information about. The name is case sensitive.

Request Lenses

getVocabulary_vocabularyName :: Lens' GetVocabulary Text

The name of the vocabulary to return information about. The name is case sensitive.

Destructuring the Response

data GetVocabularyResponse

See: newGetVocabularyResponse smart constructor.

Constructors

GetVocabularyResponse'

Fields

failureReason :: Maybe Text

If the VocabularyState field is FAILED, this field contains information about why the job failed.

languageCode :: Maybe LanguageCode

The language code of the vocabulary entries.

downloadUri :: Maybe Text

The S3 location where the vocabulary is stored. Use this URI to get the contents of the vocabulary. The URI is available for a limited time.

vocabularyName :: Maybe Text

The name of the vocabulary to return.

lastModifiedTime :: Maybe POSIX

The date and time that the vocabulary was last modified.

vocabularyState :: Maybe VocabularyState

The processing state of the vocabulary.

httpStatus :: Int

The response's http status code.

newGetVocabularyResponse

Arguments

:: Int
$sel:httpStatus:GetVocabularyResponse'

-> GetVocabularyResponse

Create a value of GetVocabularyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:GetVocabularyResponse', getVocabularyResponse_failureReason - If the VocabularyState field is FAILED, this field contains information about why the job failed.

$sel:languageCode:GetVocabularyResponse', getVocabularyResponse_languageCode - The language code of the vocabulary entries.

$sel:downloadUri:GetVocabularyResponse', getVocabularyResponse_downloadUri - The S3 location where the vocabulary is stored. Use this URI to get the contents of the vocabulary. The URI is available for a limited time.

$sel:vocabularyName:GetVocabulary', getVocabularyResponse_vocabularyName - The name of the vocabulary to return.

$sel:lastModifiedTime:GetVocabularyResponse', getVocabularyResponse_lastModifiedTime - The date and time that the vocabulary was last modified.

$sel:vocabularyState:GetVocabularyResponse', getVocabularyResponse_vocabularyState - The processing state of the vocabulary.

$sel:httpStatus:GetVocabularyResponse', getVocabularyResponse_httpStatus - The response's http status code.

Response Lenses

getVocabularyResponse_failureReason :: Lens' GetVocabularyResponse (Maybe Text)

If the VocabularyState field is FAILED, this field contains information about why the job failed.

getVocabularyResponse_languageCode :: Lens' GetVocabularyResponse (Maybe LanguageCode)

The language code of the vocabulary entries.

getVocabularyResponse_downloadUri :: Lens' GetVocabularyResponse (Maybe Text)

The S3 location where the vocabulary is stored. Use this URI to get the contents of the vocabulary. The URI is available for a limited time.

getVocabularyResponse_vocabularyName :: Lens' GetVocabularyResponse (Maybe Text)

The name of the vocabulary to return.

getVocabularyResponse_lastModifiedTime :: Lens' GetVocabularyResponse (Maybe UTCTime)

The date and time that the vocabulary was last modified.

getVocabularyResponse_vocabularyState :: Lens' GetVocabularyResponse (Maybe VocabularyState)

The processing state of the vocabulary.

getVocabularyResponse_httpStatus :: Lens' GetVocabularyResponse Int

The response's http status code.