Amazonka.Transcribe.Types.InterruptionFilter

Description

Documentation

data InterruptionFilter

An object that enables you to configure your category to be applied to call analytics jobs where either the customer or agent was interrupted.

See: newInterruptionFilter smart constructor.

Constructors

InterruptionFilter'

Fields

participantRole :: Maybe ParticipantRole

Indicates whether the caller or customer was interrupting.

relativeTimeRange :: Maybe RelativeTimeRange

An object that allows percentages to specify the proportion of the call where there was a interruption. For example, you can specify the first half of the call. You can also specify the period of time between halfway through to three-quarters of the way through the call. Because the length of conversation can vary between calls, you can apply relative time ranges across all calls.

negate :: Maybe Bool

Set to TRUE to look for a time period where there was no interruption.

threshold :: Maybe Natural

The duration of the interruption.

absoluteTimeRange :: Maybe AbsoluteTimeRange

An object you can use to specify a time range (in milliseconds) for when you'd want to find the interruption. For example, you could search for an interruption between the 30,000 millisecond mark and the 45,000 millisecond mark. You could also specify the time period as the first 15,000 milliseconds or the last 15,000 milliseconds.

newInterruptionFilter :: InterruptionFilter

Create a value of InterruptionFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:participantRole:InterruptionFilter', interruptionFilter_participantRole - Indicates whether the caller or customer was interrupting.

$sel:relativeTimeRange:InterruptionFilter', interruptionFilter_relativeTimeRange - An object that allows percentages to specify the proportion of the call where there was a interruption. For example, you can specify the first half of the call. You can also specify the period of time between halfway through to three-quarters of the way through the call. Because the length of conversation can vary between calls, you can apply relative time ranges across all calls.

$sel:negate:InterruptionFilter', interruptionFilter_negate - Set to TRUE to look for a time period where there was no interruption.

$sel:threshold:InterruptionFilter', interruptionFilter_threshold - The duration of the interruption.

$sel:absoluteTimeRange:InterruptionFilter', interruptionFilter_absoluteTimeRange - An object you can use to specify a time range (in milliseconds) for when you'd want to find the interruption. For example, you could search for an interruption between the 30,000 millisecond mark and the 45,000 millisecond mark. You could also specify the time period as the first 15,000 milliseconds or the last 15,000 milliseconds.

interruptionFilter_participantRole :: Lens' InterruptionFilter (Maybe ParticipantRole)

Indicates whether the caller or customer was interrupting.

interruptionFilter_relativeTimeRange :: Lens' InterruptionFilter (Maybe RelativeTimeRange)

An object that allows percentages to specify the proportion of the call where there was a interruption. For example, you can specify the first half of the call. You can also specify the period of time between halfway through to three-quarters of the way through the call. Because the length of conversation can vary between calls, you can apply relative time ranges across all calls.

interruptionFilter_negate :: Lens' InterruptionFilter (Maybe Bool)

Set to TRUE to look for a time period where there was no interruption.

interruptionFilter_threshold :: Lens' InterruptionFilter (Maybe Natural)

The duration of the interruption.

interruptionFilter_absoluteTimeRange :: Lens' InterruptionFilter (Maybe AbsoluteTimeRange)

An object you can use to specify a time range (in milliseconds) for when you'd want to find the interruption. For example, you could search for an interruption between the 30,000 millisecond mark and the 45,000 millisecond mark. You could also specify the time period as the first 15,000 milliseconds or the last 15,000 milliseconds.