Amazonka.Transcribe.Types.LanguageModel

Description

Documentation

data LanguageModel

The structure used to describe a custom language model.

See: newLanguageModel smart constructor.

Constructors

LanguageModel'

Fields

failureReason :: Maybe Text

The reason why the custom language model couldn't be created.

languageCode :: Maybe CLMLanguageCode

The language code you used to create your custom language model.

modelName :: Maybe Text

The name of the custom language model.

lastModifiedTime :: Maybe POSIX

The most recent time the custom language model was modified.

upgradeAvailability :: Maybe Bool

Whether the base model used for the custom language model is up to date. If this field is true then you are running the most up-to-date version of the base model in your custom language model.

inputDataConfig :: Maybe InputDataConfig

The data access role and Amazon S3 prefixes for the input files used to train the custom language model.

baseModelName :: Maybe BaseModelName

The Amazon Transcribe standard language model, or base model used to create the custom language model.

modelStatus :: Maybe ModelStatus

The creation status of a custom language model. When the status is COMPLETED the model is ready for use.

createTime :: Maybe POSIX

The time the custom language model was created.

newLanguageModel :: LanguageModel

Create a value of LanguageModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:LanguageModel', languageModel_failureReason - The reason why the custom language model couldn't be created.

$sel:languageCode:LanguageModel', languageModel_languageCode - The language code you used to create your custom language model.

$sel:modelName:LanguageModel', languageModel_modelName - The name of the custom language model.

$sel:lastModifiedTime:LanguageModel', languageModel_lastModifiedTime - The most recent time the custom language model was modified.

$sel:upgradeAvailability:LanguageModel', languageModel_upgradeAvailability - Whether the base model used for the custom language model is up to date. If this field is true then you are running the most up-to-date version of the base model in your custom language model.

$sel:inputDataConfig:LanguageModel', languageModel_inputDataConfig - The data access role and Amazon S3 prefixes for the input files used to train the custom language model.

$sel:baseModelName:LanguageModel', languageModel_baseModelName - The Amazon Transcribe standard language model, or base model used to create the custom language model.

$sel:modelStatus:LanguageModel', languageModel_modelStatus - The creation status of a custom language model. When the status is COMPLETED the model is ready for use.

$sel:createTime:LanguageModel', languageModel_createTime - The time the custom language model was created.

languageModel_failureReason :: Lens' LanguageModel (Maybe Text)

The reason why the custom language model couldn't be created.

languageModel_languageCode :: Lens' LanguageModel (Maybe CLMLanguageCode)

The language code you used to create your custom language model.

languageModel_modelName :: Lens' LanguageModel (Maybe Text)

The name of the custom language model.

languageModel_lastModifiedTime :: Lens' LanguageModel (Maybe UTCTime)

The most recent time the custom language model was modified.

languageModel_upgradeAvailability :: Lens' LanguageModel (Maybe Bool)

Whether the base model used for the custom language model is up to date. If this field is true then you are running the most up-to-date version of the base model in your custom language model.

languageModel_inputDataConfig :: Lens' LanguageModel (Maybe InputDataConfig)

The data access role and Amazon S3 prefixes for the input files used to train the custom language model.

languageModel_baseModelName :: Lens' LanguageModel (Maybe BaseModelName)

The Amazon Transcribe standard language model, or base model used to create the custom language model.

languageModel_modelStatus :: Lens' LanguageModel (Maybe ModelStatus)

The creation status of a custom language model. When the status is COMPLETED the model is ready for use.

languageModel_createTime :: Lens' LanguageModel (Maybe UTCTime)

The time the custom language model was created.