Amazonka.Transcribe.Types.Settings

Description

Documentation

data Settings

Provides optional settings for the StartTranscriptionJob operation.

See: newSettings smart constructor.

Constructors

Settings'

Fields

vocabularyName :: Maybe Text

The name of a vocabulary to use when processing the transcription job.

maxAlternatives :: Maybe Natural

The number of alternative transcriptions that the service should return. If you specify the MaxAlternatives field, you must set the ShowAlternatives field to true.

channelIdentification :: Maybe Bool

Instructs Amazon Transcribe to process each audio channel separately and then merge the transcription output of each channel into a single transcription.

Amazon Transcribe also produces a transcription of each item detected on an audio channel, including the start time and end time of the item and alternative transcriptions of the item including the confidence that Amazon Transcribe has in the transcription.

You can't set both ShowSpeakerLabels and ChannelIdentification in the same request. If you set both, your request returns a BadRequestException.

showAlternatives :: Maybe Bool

Determines whether the transcription contains alternative transcriptions. If you set the ShowAlternatives field to true, you must also set the maximum number of alternatives to return in the MaxAlternatives field.

maxSpeakerLabels :: Maybe Natural

The maximum number of speakers to identify in the input audio. If there are more speakers in the audio than this number, multiple speakers are identified as a single speaker. If you specify the MaxSpeakerLabels field, you must set the ShowSpeakerLabels field to true.

vocabularyFilterName :: Maybe Text

The name of the vocabulary filter to use when transcribing the audio. The filter that you specify must have the same language code as the transcription job.

showSpeakerLabels :: Maybe Bool

Determines whether the transcription job uses speaker recognition to identify different speakers in the input audio. Speaker recognition labels individual speakers in the audio file. If you set the ShowSpeakerLabels field to true, you must also set the maximum number of speaker labels MaxSpeakerLabels field.

You can't set both ShowSpeakerLabels and ChannelIdentification in the same request. If you set both, your request returns a BadRequestException.

vocabularyFilterMethod :: Maybe VocabularyFilterMethod

Set to mask to remove filtered text from the transcript and replace it with three asterisks ("***") as placeholder text. Set to remove to remove filtered text from the transcript without using placeholder text. Set to tag to mark the word in the transcription output that matches the vocabulary filter. When you set the filter method to tag, the words matching your vocabulary filter are not masked or removed.

newSettings :: Settings

Create a value of Settings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vocabularyName:Settings', settings_vocabularyName - The name of a vocabulary to use when processing the transcription job.

$sel:maxAlternatives:Settings', settings_maxAlternatives - The number of alternative transcriptions that the service should return. If you specify the MaxAlternatives field, you must set the ShowAlternatives field to true.

$sel:channelIdentification:Settings', settings_channelIdentification - Instructs Amazon Transcribe to process each audio channel separately and then merge the transcription output of each channel into a single transcription.

Amazon Transcribe also produces a transcription of each item detected on an audio channel, including the start time and end time of the item and alternative transcriptions of the item including the confidence that Amazon Transcribe has in the transcription.

You can't set both ShowSpeakerLabels and ChannelIdentification in the same request. If you set both, your request returns a BadRequestException.

$sel:showAlternatives:Settings', settings_showAlternatives - Determines whether the transcription contains alternative transcriptions. If you set the ShowAlternatives field to true, you must also set the maximum number of alternatives to return in the MaxAlternatives field.

$sel:maxSpeakerLabels:Settings', settings_maxSpeakerLabels - The maximum number of speakers to identify in the input audio. If there are more speakers in the audio than this number, multiple speakers are identified as a single speaker. If you specify the MaxSpeakerLabels field, you must set the ShowSpeakerLabels field to true.

$sel:vocabularyFilterName:Settings', settings_vocabularyFilterName - The name of the vocabulary filter to use when transcribing the audio. The filter that you specify must have the same language code as the transcription job.

$sel:showSpeakerLabels:Settings', settings_showSpeakerLabels - Determines whether the transcription job uses speaker recognition to identify different speakers in the input audio. Speaker recognition labels individual speakers in the audio file. If you set the ShowSpeakerLabels field to true, you must also set the maximum number of speaker labels MaxSpeakerLabels field.

You can't set both ShowSpeakerLabels and ChannelIdentification in the same request. If you set both, your request returns a BadRequestException.

$sel:vocabularyFilterMethod:Settings', settings_vocabularyFilterMethod - Set to mask to remove filtered text from the transcript and replace it with three asterisks ("***") as placeholder text. Set to remove to remove filtered text from the transcript without using placeholder text. Set to tag to mark the word in the transcription output that matches the vocabulary filter. When you set the filter method to tag, the words matching your vocabulary filter are not masked or removed.

settings_vocabularyName :: Lens' Settings (Maybe Text)

The name of a vocabulary to use when processing the transcription job.

settings_maxAlternatives :: Lens' Settings (Maybe Natural)

The number of alternative transcriptions that the service should return. If you specify the MaxAlternatives field, you must set the ShowAlternatives field to true.

settings_channelIdentification :: Lens' Settings (Maybe Bool)

Instructs Amazon Transcribe to process each audio channel separately and then merge the transcription output of each channel into a single transcription.

Amazon Transcribe also produces a transcription of each item detected on an audio channel, including the start time and end time of the item and alternative transcriptions of the item including the confidence that Amazon Transcribe has in the transcription.

You can't set both ShowSpeakerLabels and ChannelIdentification in the same request. If you set both, your request returns a BadRequestException.

settings_showAlternatives :: Lens' Settings (Maybe Bool)

Determines whether the transcription contains alternative transcriptions. If you set the ShowAlternatives field to true, you must also set the maximum number of alternatives to return in the MaxAlternatives field.

settings_maxSpeakerLabels :: Lens' Settings (Maybe Natural)

The maximum number of speakers to identify in the input audio. If there are more speakers in the audio than this number, multiple speakers are identified as a single speaker. If you specify the MaxSpeakerLabels field, you must set the ShowSpeakerLabels field to true.

settings_vocabularyFilterName :: Lens' Settings (Maybe Text)

The name of the vocabulary filter to use when transcribing the audio. The filter that you specify must have the same language code as the transcription job.

settings_showSpeakerLabels :: Lens' Settings (Maybe Bool)

Determines whether the transcription job uses speaker recognition to identify different speakers in the input audio. Speaker recognition labels individual speakers in the audio file. If you set the ShowSpeakerLabels field to true, you must also set the maximum number of speaker labels MaxSpeakerLabels field.

You can't set both ShowSpeakerLabels and ChannelIdentification in the same request. If you set both, your request returns a BadRequestException.

settings_vocabularyFilterMethod :: Lens' Settings (Maybe VocabularyFilterMethod)

Set to mask to remove filtered text from the transcript and replace it with three asterisks ("***") as placeholder text. Set to remove to remove filtered text from the transcript without using placeholder text. Set to tag to mark the word in the transcription output that matches the vocabulary filter. When you set the filter method to tag, the words matching your vocabulary filter are not masked or removed.