Amazonka.Transcribe.Types.TranscriptionJob

Description

Documentation

data TranscriptionJob

Describes an asynchronous transcription job that was created with the StartTranscriptionJob operation.

See: newTranscriptionJob smart constructor.

Constructors

TranscriptionJob'

Fields

creationTime :: Maybe POSIX

A timestamp that shows when the job was created.

failureReason :: Maybe Text

If the TranscriptionJobStatus field is FAILED, this field contains information about why the job failed.

The FailureReason field can contain one of the following values:

Unsupported media format - The media format specified in the MediaFormat field of the request isn't valid. See the description of the MediaFormat field for a list of valid values.

The media format provided does not match the detected media format

The media format of the audio file doesn't match the format specified in the MediaFormat field in the request. Check the media format of your media file and make sure that the two values match.

Invalid sample rate for audio file - The sample rate specified in the MediaSampleRateHertz of the request isn't valid. The sample rate must be between 8,000 and 48,000 Hertz.

The sample rate provided does not match the detected sample rate - The sample rate in the audio file doesn't match the sample rate specified in the MediaSampleRateHertz field in the request. Check the sample rate of your media file and make sure that the two values match.

Invalid file size: file size too large - The size of your audio file is larger than Amazon Transcribe can process. For more information, see Limits in the Amazon Transcribe Developer Guide.

Invalid number of channels: number of channels too large - Your audio contains more channels than Amazon Transcribe is configured to process. To request additional channels, see Amazon Transcribe Limits in the Amazon Web Services General Reference.

contentRedaction :: Maybe ContentRedaction

An object that describes content redaction settings for the transcription job.

identifiedLanguageScore :: Maybe Double

A value between zero and one that Amazon Transcribe assigned to the language that it identified in the source audio. Larger values indicate that Amazon Transcribe has higher confidence in the language it identified.

subtitles :: Maybe SubtitlesOutput

Generate subtitles for your batch transcription job.

languageCode :: Maybe LanguageCode

The language code for the input speech.

languageOptions :: Maybe (NonEmpty LanguageCode)

An object that shows the optional array of languages inputted for transcription jobs with automatic language identification enabled.

settings :: Maybe Settings

Optional settings for the transcription job. Use these settings to turn on speaker recognition, to set the maximum number of speakers that should be identified and to specify a custom vocabulary to use when processing the transcription job.

startTime :: Maybe POSIX

A timestamp that shows when the job started processing.

completionTime :: Maybe POSIX

A timestamp that shows when the job completed.

media :: Maybe Media

An object that describes the input media for the transcription job.

mediaFormat :: Maybe MediaFormat

The format of the input media file.

modelSettings :: Maybe ModelSettings

An object containing the details of your custom language model.

transcriptionJobStatus :: Maybe TranscriptionJobStatus

The status of the transcription job.

jobExecutionSettings :: Maybe JobExecutionSettings

Provides information about how a transcription job is executed.

transcriptionJobName :: Maybe Text

The name of the transcription job.

identifyLanguage :: Maybe Bool

A value that shows if automatic language identification was enabled for a transcription job.

transcript :: Maybe Transcript

An object that describes the output of the transcription job.

tags :: Maybe (NonEmpty Tag)

A key:value pair assigned to a given transcription job.

mediaSampleRateHertz :: Maybe Natural

The sample rate, in Hertz, of the audio track in the input media file.

newTranscriptionJob :: TranscriptionJob

Create a value of TranscriptionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:TranscriptionJob', transcriptionJob_creationTime - A timestamp that shows when the job was created.

$sel:failureReason:TranscriptionJob', transcriptionJob_failureReason - If the TranscriptionJobStatus field is FAILED, this field contains information about why the job failed.

The FailureReason field can contain one of the following values:

Unsupported media format - The media format specified in the MediaFormat field of the request isn't valid. See the description of the MediaFormat field for a list of valid values.

The media format provided does not match the detected media format

The media format of the audio file doesn't match the format specified in the MediaFormat field in the request. Check the media format of your media file and make sure that the two values match.

Invalid sample rate for audio file - The sample rate specified in the MediaSampleRateHertz of the request isn't valid. The sample rate must be between 8,000 and 48,000 Hertz.

The sample rate provided does not match the detected sample rate - The sample rate in the audio file doesn't match the sample rate specified in the MediaSampleRateHertz field in the request. Check the sample rate of your media file and make sure that the two values match.

Invalid file size: file size too large - The size of your audio file is larger than Amazon Transcribe can process. For more information, see Limits in the Amazon Transcribe Developer Guide.

Invalid number of channels: number of channels too large - Your audio contains more channels than Amazon Transcribe is configured to process. To request additional channels, see Amazon Transcribe Limits in the Amazon Web Services General Reference.

$sel:contentRedaction:TranscriptionJob', transcriptionJob_contentRedaction - An object that describes content redaction settings for the transcription job.

$sel:identifiedLanguageScore:TranscriptionJob', transcriptionJob_identifiedLanguageScore - A value between zero and one that Amazon Transcribe assigned to the language that it identified in the source audio. Larger values indicate that Amazon Transcribe has higher confidence in the language it identified.

$sel:subtitles:TranscriptionJob', transcriptionJob_subtitles - Generate subtitles for your batch transcription job.

$sel:languageCode:TranscriptionJob', transcriptionJob_languageCode - The language code for the input speech.

$sel:languageOptions:TranscriptionJob', transcriptionJob_languageOptions - An object that shows the optional array of languages inputted for transcription jobs with automatic language identification enabled.

$sel:settings:TranscriptionJob', transcriptionJob_settings - Optional settings for the transcription job. Use these settings to turn on speaker recognition, to set the maximum number of speakers that should be identified and to specify a custom vocabulary to use when processing the transcription job.

$sel:startTime:TranscriptionJob', transcriptionJob_startTime - A timestamp that shows when the job started processing.

$sel:completionTime:TranscriptionJob', transcriptionJob_completionTime - A timestamp that shows when the job completed.

$sel:media:TranscriptionJob', transcriptionJob_media - An object that describes the input media for the transcription job.

$sel:mediaFormat:TranscriptionJob', transcriptionJob_mediaFormat - The format of the input media file.

$sel:modelSettings:TranscriptionJob', transcriptionJob_modelSettings - An object containing the details of your custom language model.

$sel:transcriptionJobStatus:TranscriptionJob', transcriptionJob_transcriptionJobStatus - The status of the transcription job.

$sel:jobExecutionSettings:TranscriptionJob', transcriptionJob_jobExecutionSettings - Provides information about how a transcription job is executed.

$sel:transcriptionJobName:TranscriptionJob', transcriptionJob_transcriptionJobName - The name of the transcription job.

$sel:identifyLanguage:TranscriptionJob', transcriptionJob_identifyLanguage - A value that shows if automatic language identification was enabled for a transcription job.

$sel:transcript:TranscriptionJob', transcriptionJob_transcript - An object that describes the output of the transcription job.

$sel:tags:TranscriptionJob', transcriptionJob_tags - A key:value pair assigned to a given transcription job.

$sel:mediaSampleRateHertz:TranscriptionJob', transcriptionJob_mediaSampleRateHertz - The sample rate, in Hertz, of the audio track in the input media file.

transcriptionJob_creationTime :: Lens' TranscriptionJob (Maybe UTCTime)

A timestamp that shows when the job was created.

transcriptionJob_failureReason :: Lens' TranscriptionJob (Maybe Text)

If the TranscriptionJobStatus field is FAILED, this field contains information about why the job failed.

The FailureReason field can contain one of the following values:

Unsupported media format - The media format specified in the MediaFormat field of the request isn't valid. See the description of the MediaFormat field for a list of valid values.

The media format provided does not match the detected media format

The media format of the audio file doesn't match the format specified in the MediaFormat field in the request. Check the media format of your media file and make sure that the two values match.

Invalid sample rate for audio file - The sample rate specified in the MediaSampleRateHertz of the request isn't valid. The sample rate must be between 8,000 and 48,000 Hertz.

The sample rate provided does not match the detected sample rate - The sample rate in the audio file doesn't match the sample rate specified in the MediaSampleRateHertz field in the request. Check the sample rate of your media file and make sure that the two values match.

Invalid file size: file size too large - The size of your audio file is larger than Amazon Transcribe can process. For more information, see Limits in the Amazon Transcribe Developer Guide.

Invalid number of channels: number of channels too large - Your audio contains more channels than Amazon Transcribe is configured to process. To request additional channels, see Amazon Transcribe Limits in the Amazon Web Services General Reference.

transcriptionJob_contentRedaction :: Lens' TranscriptionJob (Maybe ContentRedaction)

An object that describes content redaction settings for the transcription job.

transcriptionJob_identifiedLanguageScore :: Lens' TranscriptionJob (Maybe Double)

A value between zero and one that Amazon Transcribe assigned to the language that it identified in the source audio. Larger values indicate that Amazon Transcribe has higher confidence in the language it identified.

transcriptionJob_subtitles :: Lens' TranscriptionJob (Maybe SubtitlesOutput)

Generate subtitles for your batch transcription job.

transcriptionJob_languageCode :: Lens' TranscriptionJob (Maybe LanguageCode)

The language code for the input speech.

transcriptionJob_languageOptions :: Lens' TranscriptionJob (Maybe (NonEmpty LanguageCode))

An object that shows the optional array of languages inputted for transcription jobs with automatic language identification enabled.

transcriptionJob_settings :: Lens' TranscriptionJob (Maybe Settings)

Optional settings for the transcription job. Use these settings to turn on speaker recognition, to set the maximum number of speakers that should be identified and to specify a custom vocabulary to use when processing the transcription job.

transcriptionJob_startTime :: Lens' TranscriptionJob (Maybe UTCTime)

A timestamp that shows when the job started processing.

transcriptionJob_completionTime :: Lens' TranscriptionJob (Maybe UTCTime)

A timestamp that shows when the job completed.

transcriptionJob_media :: Lens' TranscriptionJob (Maybe Media)

An object that describes the input media for the transcription job.

transcriptionJob_mediaFormat :: Lens' TranscriptionJob (Maybe MediaFormat)

The format of the input media file.

transcriptionJob_modelSettings :: Lens' TranscriptionJob (Maybe ModelSettings)

An object containing the details of your custom language model.

transcriptionJob_transcriptionJobStatus :: Lens' TranscriptionJob (Maybe TranscriptionJobStatus)

The status of the transcription job.

transcriptionJob_jobExecutionSettings :: Lens' TranscriptionJob (Maybe JobExecutionSettings)

Provides information about how a transcription job is executed.

transcriptionJob_transcriptionJobName :: Lens' TranscriptionJob (Maybe Text)

The name of the transcription job.

transcriptionJob_identifyLanguage :: Lens' TranscriptionJob (Maybe Bool)

A value that shows if automatic language identification was enabled for a transcription job.

transcriptionJob_transcript :: Lens' TranscriptionJob (Maybe Transcript)

An object that describes the output of the transcription job.

transcriptionJob_tags :: Lens' TranscriptionJob (Maybe (NonEmpty Tag))

A key:value pair assigned to a given transcription job.

transcriptionJob_mediaSampleRateHertz :: Lens' TranscriptionJob (Maybe Natural)

The sample rate, in Hertz, of the audio track in the input media file.