Amazonka.Transcribe.UpdateVocabulary

Description

Updates an existing vocabulary with new values. The UpdateVocabulary operation overwrites all of the existing information with the values that you provide in the request.

Creating a Request

data UpdateVocabulary

See: newUpdateVocabulary smart constructor.

Constructors

UpdateVocabulary'

Fields

vocabularyFileUri :: Maybe Text

The S3 location of the text file that contains the definition of the custom vocabulary. The URI must be in the same region as the API endpoint that you are calling. The general form is

For example:

For more information about S3 object names, see Object Keys in the Amazon S3 Developer Guide.

For more information about custom vocabularies, see Custom Vocabularies.

phrases :: Maybe [Text]

An array of strings containing the vocabulary entries.

vocabularyName :: Text

The name of the vocabulary to update. The name is case sensitive. If you try to update a vocabulary with the same name as a previous vocabulary you will receive a ConflictException error.

languageCode :: LanguageCode

The language code of the vocabulary entries. For a list of languages and their corresponding language codes, see transcribe-whatis.

newUpdateVocabulary

Arguments

:: Text
$sel:vocabularyName:UpdateVocabulary'

-> LanguageCode
$sel:languageCode:UpdateVocabulary'

-> UpdateVocabulary

Create a value of UpdateVocabulary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vocabularyFileUri:UpdateVocabulary', updateVocabulary_vocabularyFileUri - The S3 location of the text file that contains the definition of the custom vocabulary. The URI must be in the same region as the API endpoint that you are calling. The general form is

For example:

For more information about S3 object names, see Object Keys in the Amazon S3 Developer Guide.

For more information about custom vocabularies, see Custom Vocabularies.

$sel:phrases:UpdateVocabulary', updateVocabulary_phrases - An array of strings containing the vocabulary entries.

$sel:vocabularyName:UpdateVocabulary', updateVocabulary_vocabularyName - The name of the vocabulary to update. The name is case sensitive. If you try to update a vocabulary with the same name as a previous vocabulary you will receive a ConflictException error.

$sel:languageCode:UpdateVocabulary', updateVocabulary_languageCode - The language code of the vocabulary entries. For a list of languages and their corresponding language codes, see transcribe-whatis.

Request Lenses

updateVocabulary_vocabularyFileUri :: Lens' UpdateVocabulary (Maybe Text)

The S3 location of the text file that contains the definition of the custom vocabulary. The URI must be in the same region as the API endpoint that you are calling. The general form is

For example:

For more information about S3 object names, see Object Keys in the Amazon S3 Developer Guide.

For more information about custom vocabularies, see Custom Vocabularies.

updateVocabulary_phrases :: Lens' UpdateVocabulary (Maybe [Text])

An array of strings containing the vocabulary entries.

updateVocabulary_vocabularyName :: Lens' UpdateVocabulary Text

The name of the vocabulary to update. The name is case sensitive. If you try to update a vocabulary with the same name as a previous vocabulary you will receive a ConflictException error.

updateVocabulary_languageCode :: Lens' UpdateVocabulary LanguageCode

The language code of the vocabulary entries. For a list of languages and their corresponding language codes, see transcribe-whatis.

Destructuring the Response

data UpdateVocabularyResponse

See: newUpdateVocabularyResponse smart constructor.

Constructors

UpdateVocabularyResponse'

Fields

languageCode :: Maybe LanguageCode

The language code of the vocabulary entries.

vocabularyName :: Maybe Text

The name of the vocabulary that was updated.

lastModifiedTime :: Maybe POSIX

The date and time that the vocabulary was updated.

vocabularyState :: Maybe VocabularyState

The processing state of the vocabulary. When the VocabularyState field contains READY the vocabulary is ready to be used in a StartTranscriptionJob request.

httpStatus :: Int

The response's http status code.

newUpdateVocabularyResponse

Arguments

:: Int
$sel:httpStatus:UpdateVocabularyResponse'

-> UpdateVocabularyResponse

Create a value of UpdateVocabularyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:UpdateVocabulary', updateVocabularyResponse_languageCode - The language code of the vocabulary entries.

$sel:vocabularyName:UpdateVocabulary', updateVocabularyResponse_vocabularyName - The name of the vocabulary that was updated.

$sel:lastModifiedTime:UpdateVocabularyResponse', updateVocabularyResponse_lastModifiedTime - The date and time that the vocabulary was updated.

$sel:vocabularyState:UpdateVocabularyResponse', updateVocabularyResponse_vocabularyState - The processing state of the vocabulary. When the VocabularyState field contains READY the vocabulary is ready to be used in a StartTranscriptionJob request.

$sel:httpStatus:UpdateVocabularyResponse', updateVocabularyResponse_httpStatus - The response's http status code.

Response Lenses

updateVocabularyResponse_languageCode :: Lens' UpdateVocabularyResponse (Maybe LanguageCode)

The language code of the vocabulary entries.

updateVocabularyResponse_vocabularyName :: Lens' UpdateVocabularyResponse (Maybe Text)

The name of the vocabulary that was updated.

updateVocabularyResponse_lastModifiedTime :: Lens' UpdateVocabularyResponse (Maybe UTCTime)

The date and time that the vocabulary was updated.

updateVocabularyResponse_vocabularyState :: Lens' UpdateVocabularyResponse (Maybe VocabularyState)

The processing state of the vocabulary. When the VocabularyState field contains READY the vocabulary is ready to be used in a StartTranscriptionJob request.

updateVocabularyResponse_httpStatus :: Lens' UpdateVocabularyResponse Int

The response's http status code.