Amazonka.Transfer.CreateUser

Description

Creates a user and associates them with an existing file transfer protocol-enabled server. You can only create and associate users with servers that have the IdentityProviderType set to SERVICE_MANAGED. Using parameters for CreateUser, you can specify the user name, set the home directory, store the user's public key, and assign the user's Amazon Web Services Identity and Access Management (IAM) role. You can also optionally add a session policy, and assign metadata with tags that can be used to group and search for users.

Creating a Request

data CreateUser

See: newCreateUser smart constructor.

Constructors

CreateUser'

Fields

homeDirectoryType :: Maybe HomeDirectoryType

The type of landing directory (folder) you want your users' home directory to be when they log into the server. If you set it to PATH, the user will see the absolute Amazon S3 bucket or EFS paths as is in their file transfer protocol clients. If you set it LOGICAL, you need to provide mappings in the HomeDirectoryMappings for how you want to make Amazon S3 or EFS paths visible to your users.

sshPublicKeyBody :: Maybe Text

The public portion of the Secure Shell (SSH) key used to authenticate the user to the server.

posixProfile :: Maybe PosixProfile

Specifies the full POSIX identity, including user ID (Uid), group ID (Gid), and any secondary groups IDs (SecondaryGids), that controls your users' access to your Amazon EFS file systems. The POSIX permissions that are set on files and directories in Amazon EFS determine the level of access your users get when transferring files into and out of your Amazon EFS file systems.

homeDirectoryMappings :: Maybe (NonEmpty HomeDirectoryMapEntry)

Logical directory mappings that specify what Amazon S3 or Amazon EFS paths and keys should be visible to your user and how you want to make them visible. You must specify the Entry and Target pair, where Entry shows how the path is made visible and Target is the actual Amazon S3 or Amazon EFS path. If you only specify a target, it is displayed as is. You also must ensure that your Amazon Web Services Identity and Access Management (IAM) role provides access to paths in Target. This value can only be set when HomeDirectoryType is set to LOGICAL.

The following is an Entry and Target pair example.

[ { "Entry": "/directory1", "Target": "/bucket_name/home/mydirectory" } ]

In most cases, you can use this value instead of the session policy to lock your user down to the designated home directory ("chroot"). To do this, you can set Entry to / and set Target to the HomeDirectory parameter value.

The following is an Entry and Target pair example for chroot.

[ { "Entry:": "/", "Target": "/bucket_name/home/mydirectory" } ]

If the target of a logical directory entry does not exist in Amazon S3 or EFS, the entry is ignored. As a workaround, you can use the Amazon S3 API or EFS API to create 0 byte objects as place holders for your directory. If using the CLI, use the s3api or efsapi call instead of s3 or efs so you can use the put-object operation. For example, you use the following: aws s3api put-object --bucket bucketname --key path/to/folder/. Make sure that the end of the key name ends in a / for it to be considered a folder.

policy :: Maybe Text

A session policy for your user so that you can use the same IAM role across multiple users. This policy scopes down user access to portions of their Amazon S3 bucket. Variables that you can use inside this policy include ${Transfer:UserName}, ${Transfer:HomeDirectory}, and ${Transfer:HomeBucket}.

This only applies when the domain of ServerId is S3. EFS does not use session policies.

For session policies, Amazon Web Services Transfer Family stores the policy as a JSON blob, instead of the Amazon Resource Name (ARN) of the policy. You save the policy as a JSON blob and pass it in the Policy argument.

For an example of a session policy, see Example session policy.

For more information, see AssumeRole in the Amazon Web Services Security Token Service API Reference.

homeDirectory :: Maybe Text

The landing directory (folder) for a user when they log in to the server using the client.

A HomeDirectory example is /bucket_name/home/mydirectory.

tags :: Maybe (NonEmpty Tag)

Key-value pairs that can be used to group and search for users. Tags are metadata attached to users for any purpose.

role' :: Text

Specifies the Amazon Resource Name (ARN) of the IAM role that controls your users' access to your Amazon S3 bucket or EFS file system. The policies attached to this role determine the level of access that you want to provide your users when transferring files into and out of your Amazon S3 bucket or EFS file system. The IAM role should also contain a trust relationship that allows the server to access your resources when servicing your users' transfer requests.

serverId :: Text

A system-assigned unique identifier for a server instance. This is the specific server that you added your user to.

userName :: Text

A unique string that identifies a user and is associated with a ServerId. This user name must be a minimum of 3 and a maximum of 100 characters long. The following are valid characters: a-z, A-Z, 0-9, underscore '_', hyphen '-', period '.', and at sign '@'. The user name can't start with a hyphen, period, or at sign.

newCreateUser

Arguments

:: Text
$sel:role':CreateUser'

-> Text
$sel:serverId:CreateUser'

-> Text
$sel:userName:CreateUser'

-> CreateUser

Create a value of CreateUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:homeDirectoryType:CreateUser', createUser_homeDirectoryType - The type of landing directory (folder) you want your users' home directory to be when they log into the server. If you set it to PATH, the user will see the absolute Amazon S3 bucket or EFS paths as is in their file transfer protocol clients. If you set it LOGICAL, you need to provide mappings in the HomeDirectoryMappings for how you want to make Amazon S3 or EFS paths visible to your users.

$sel:sshPublicKeyBody:CreateUser', createUser_sshPublicKeyBody - The public portion of the Secure Shell (SSH) key used to authenticate the user to the server.

$sel:posixProfile:CreateUser', createUser_posixProfile - Specifies the full POSIX identity, including user ID (Uid), group ID (Gid), and any secondary groups IDs (SecondaryGids), that controls your users' access to your Amazon EFS file systems. The POSIX permissions that are set on files and directories in Amazon EFS determine the level of access your users get when transferring files into and out of your Amazon EFS file systems.

$sel:homeDirectoryMappings:CreateUser', createUser_homeDirectoryMappings - Logical directory mappings that specify what Amazon S3 or Amazon EFS paths and keys should be visible to your user and how you want to make them visible. You must specify the Entry and Target pair, where Entry shows how the path is made visible and Target is the actual Amazon S3 or Amazon EFS path. If you only specify a target, it is displayed as is. You also must ensure that your Amazon Web Services Identity and Access Management (IAM) role provides access to paths in Target. This value can only be set when HomeDirectoryType is set to LOGICAL.

The following is an Entry and Target pair example.

[ { "Entry": "/directory1", "Target": "/bucket_name/home/mydirectory" } ]

In most cases, you can use this value instead of the session policy to lock your user down to the designated home directory ("chroot"). To do this, you can set Entry to / and set Target to the HomeDirectory parameter value.

The following is an Entry and Target pair example for chroot.

[ { "Entry:": "/", "Target": "/bucket_name/home/mydirectory" } ]

If the target of a logical directory entry does not exist in Amazon S3 or EFS, the entry is ignored. As a workaround, you can use the Amazon S3 API or EFS API to create 0 byte objects as place holders for your directory. If using the CLI, use the s3api or efsapi call instead of s3 or efs so you can use the put-object operation. For example, you use the following: aws s3api put-object --bucket bucketname --key path/to/folder/. Make sure that the end of the key name ends in a / for it to be considered a folder.

$sel:policy:CreateUser', createUser_policy - A session policy for your user so that you can use the same IAM role across multiple users. This policy scopes down user access to portions of their Amazon S3 bucket. Variables that you can use inside this policy include ${Transfer:UserName}, ${Transfer:HomeDirectory}, and ${Transfer:HomeBucket}.

This only applies when the domain of ServerId is S3. EFS does not use session policies.

For session policies, Amazon Web Services Transfer Family stores the policy as a JSON blob, instead of the Amazon Resource Name (ARN) of the policy. You save the policy as a JSON blob and pass it in the Policy argument.

For an example of a session policy, see Example session policy.

For more information, see AssumeRole in the Amazon Web Services Security Token Service API Reference.

$sel:homeDirectory:CreateUser', createUser_homeDirectory - The landing directory (folder) for a user when they log in to the server using the client.

A HomeDirectory example is /bucket_name/home/mydirectory.

$sel:tags:CreateUser', createUser_tags - Key-value pairs that can be used to group and search for users. Tags are metadata attached to users for any purpose.

$sel:role':CreateUser', createUser_role - Specifies the Amazon Resource Name (ARN) of the IAM role that controls your users' access to your Amazon S3 bucket or EFS file system. The policies attached to this role determine the level of access that you want to provide your users when transferring files into and out of your Amazon S3 bucket or EFS file system. The IAM role should also contain a trust relationship that allows the server to access your resources when servicing your users' transfer requests.

$sel:serverId:CreateUser', createUser_serverId - A system-assigned unique identifier for a server instance. This is the specific server that you added your user to.

$sel:userName:CreateUser', createUser_userName - A unique string that identifies a user and is associated with a ServerId. This user name must be a minimum of 3 and a maximum of 100 characters long. The following are valid characters: a-z, A-Z, 0-9, underscore '_', hyphen '-', period '.', and at sign '@'. The user name can't start with a hyphen, period, or at sign.

Request Lenses

createUser_homeDirectoryType :: Lens' CreateUser (Maybe HomeDirectoryType)

The type of landing directory (folder) you want your users' home directory to be when they log into the server. If you set it to PATH, the user will see the absolute Amazon S3 bucket or EFS paths as is in their file transfer protocol clients. If you set it LOGICAL, you need to provide mappings in the HomeDirectoryMappings for how you want to make Amazon S3 or EFS paths visible to your users.

createUser_sshPublicKeyBody :: Lens' CreateUser (Maybe Text)

The public portion of the Secure Shell (SSH) key used to authenticate the user to the server.

createUser_posixProfile :: Lens' CreateUser (Maybe PosixProfile)

Specifies the full POSIX identity, including user ID (Uid), group ID (Gid), and any secondary groups IDs (SecondaryGids), that controls your users' access to your Amazon EFS file systems. The POSIX permissions that are set on files and directories in Amazon EFS determine the level of access your users get when transferring files into and out of your Amazon EFS file systems.

createUser_homeDirectoryMappings :: Lens' CreateUser (Maybe (NonEmpty HomeDirectoryMapEntry))

Logical directory mappings that specify what Amazon S3 or Amazon EFS paths and keys should be visible to your user and how you want to make them visible. You must specify the Entry and Target pair, where Entry shows how the path is made visible and Target is the actual Amazon S3 or Amazon EFS path. If you only specify a target, it is displayed as is. You also must ensure that your Amazon Web Services Identity and Access Management (IAM) role provides access to paths in Target. This value can only be set when HomeDirectoryType is set to LOGICAL.

The following is an Entry and Target pair example.

[ { "Entry": "/directory1", "Target": "/bucket_name/home/mydirectory" } ]

In most cases, you can use this value instead of the session policy to lock your user down to the designated home directory ("chroot"). To do this, you can set Entry to / and set Target to the HomeDirectory parameter value.

The following is an Entry and Target pair example for chroot.

[ { "Entry:": "/", "Target": "/bucket_name/home/mydirectory" } ]

If the target of a logical directory entry does not exist in Amazon S3 or EFS, the entry is ignored. As a workaround, you can use the Amazon S3 API or EFS API to create 0 byte objects as place holders for your directory. If using the CLI, use the s3api or efsapi call instead of s3 or efs so you can use the put-object operation. For example, you use the following: aws s3api put-object --bucket bucketname --key path/to/folder/. Make sure that the end of the key name ends in a / for it to be considered a folder.

createUser_policy :: Lens' CreateUser (Maybe Text)

A session policy for your user so that you can use the same IAM role across multiple users. This policy scopes down user access to portions of their Amazon S3 bucket. Variables that you can use inside this policy include ${Transfer:UserName}, ${Transfer:HomeDirectory}, and ${Transfer:HomeBucket}.

This only applies when the domain of ServerId is S3. EFS does not use session policies.

For session policies, Amazon Web Services Transfer Family stores the policy as a JSON blob, instead of the Amazon Resource Name (ARN) of the policy. You save the policy as a JSON blob and pass it in the Policy argument.

For an example of a session policy, see Example session policy.

For more information, see AssumeRole in the Amazon Web Services Security Token Service API Reference.

createUser_homeDirectory :: Lens' CreateUser (Maybe Text)

The landing directory (folder) for a user when they log in to the server using the client.

A HomeDirectory example is /bucket_name/home/mydirectory.

createUser_tags :: Lens' CreateUser (Maybe (NonEmpty Tag))

Key-value pairs that can be used to group and search for users. Tags are metadata attached to users for any purpose.

createUser_role :: Lens' CreateUser Text

Specifies the Amazon Resource Name (ARN) of the IAM role that controls your users' access to your Amazon S3 bucket or EFS file system. The policies attached to this role determine the level of access that you want to provide your users when transferring files into and out of your Amazon S3 bucket or EFS file system. The IAM role should also contain a trust relationship that allows the server to access your resources when servicing your users' transfer requests.

createUser_serverId :: Lens' CreateUser Text

A system-assigned unique identifier for a server instance. This is the specific server that you added your user to.

createUser_userName :: Lens' CreateUser Text

A unique string that identifies a user and is associated with a ServerId. This user name must be a minimum of 3 and a maximum of 100 characters long. The following are valid characters: a-z, A-Z, 0-9, underscore '_', hyphen '-', period '.', and at sign '@'. The user name can't start with a hyphen, period, or at sign.

Destructuring the Response

data CreateUserResponse

See: newCreateUserResponse smart constructor.

Constructors

CreateUserResponse'

Fields

httpStatus :: Int

The response's http status code.

serverId :: Text

The ID of the server that the user is attached to.

userName :: Text

A unique string that identifies a user account associated with a server.

newCreateUserResponse

Arguments

:: Int
$sel:httpStatus:CreateUserResponse'

-> Text
$sel:serverId:CreateUser'

-> Text
$sel:userName:CreateUser'

-> CreateUserResponse

Create a value of CreateUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateUserResponse', createUserResponse_httpStatus - The response's http status code.

$sel:serverId:CreateUser', createUserResponse_serverId - The ID of the server that the user is attached to.

$sel:userName:CreateUser', createUserResponse_userName - A unique string that identifies a user account associated with a server.

Response Lenses

createUserResponse_httpStatus :: Lens' CreateUserResponse Int

The response's http status code.

createUserResponse_serverId :: Lens' CreateUserResponse Text

The ID of the server that the user is attached to.

createUserResponse_userName :: Lens' CreateUserResponse Text

A unique string that identifies a user account associated with a server.