Amazonka.Transfer.DescribeAccess

Description

Describes the access that is assigned to the specific file transfer protocol-enabled server, as identified by its ServerId property and its ExternalID.

The response from this call returns the properties of the access that is associated with the ServerId value that was specified.

Creating a Request

data DescribeAccess

See: newDescribeAccess smart constructor.

Constructors

DescribeAccess'

Fields

serverId :: Text

A system-assigned unique identifier for a server that has this access assigned.

externalId :: Text

A unique identifier that is required to identify specific groups within your directory. The users of the group that you associate have access to your Amazon S3 or Amazon EFS resources over the enabled protocols using Amazon Web Services Transfer Family. If you know the group name, you can view the SID values by running the following command using Windows PowerShell.

Get-ADGroup -Filter {samAccountName -like "YourGroupName*"} -Properties * | Select SamAccountName,ObjectSid

In that command, replace YourGroupName with the name of your Active Directory group.

The regex used to validate this parameter is a string of characters consisting of uppercase and lowercase alphanumeric characters with no spaces. You can also include underscores or any of the following characters: =,.@:/-

newDescribeAccess

Arguments

:: Text
$sel:serverId:DescribeAccess'

-> Text
$sel:externalId:DescribeAccess'

-> DescribeAccess

Create a value of DescribeAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverId:DescribeAccess', describeAccess_serverId - A system-assigned unique identifier for a server that has this access assigned.

$sel:externalId:DescribeAccess', describeAccess_externalId - A unique identifier that is required to identify specific groups within your directory. The users of the group that you associate have access to your Amazon S3 or Amazon EFS resources over the enabled protocols using Amazon Web Services Transfer Family. If you know the group name, you can view the SID values by running the following command using Windows PowerShell.

Get-ADGroup -Filter {samAccountName -like "YourGroupName*"} -Properties * | Select SamAccountName,ObjectSid

In that command, replace YourGroupName with the name of your Active Directory group.

The regex used to validate this parameter is a string of characters consisting of uppercase and lowercase alphanumeric characters with no spaces. You can also include underscores or any of the following characters: =,.@:/-

Request Lenses

describeAccess_serverId :: Lens' DescribeAccess Text

A system-assigned unique identifier for a server that has this access assigned.

describeAccess_externalId :: Lens' DescribeAccess Text

A unique identifier that is required to identify specific groups within your directory. The users of the group that you associate have access to your Amazon S3 or Amazon EFS resources over the enabled protocols using Amazon Web Services Transfer Family. If you know the group name, you can view the SID values by running the following command using Windows PowerShell.

Get-ADGroup -Filter {samAccountName -like "YourGroupName*"} -Properties * | Select SamAccountName,ObjectSid

In that command, replace YourGroupName with the name of your Active Directory group.

The regex used to validate this parameter is a string of characters consisting of uppercase and lowercase alphanumeric characters with no spaces. You can also include underscores or any of the following characters: =,.@:/-

Destructuring the Response

data DescribeAccessResponse

See: newDescribeAccessResponse smart constructor.

Constructors

DescribeAccessResponse'

Fields

httpStatus :: Int

The response's http status code.

serverId :: Text

A system-assigned unique identifier for a server that has this access assigned.

access :: DescribedAccess

The external ID of the server that the access is attached to.

newDescribeAccessResponse

Arguments

:: Int
$sel:httpStatus:DescribeAccessResponse'

-> Text
$sel:serverId:DescribeAccess'

-> DescribedAccess
$sel:access:DescribeAccessResponse'

-> DescribeAccessResponse

Create a value of DescribeAccessResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeAccessResponse', describeAccessResponse_httpStatus - The response's http status code.

$sel:serverId:DescribeAccess', describeAccessResponse_serverId - A system-assigned unique identifier for a server that has this access assigned.

$sel:access:DescribeAccessResponse', describeAccessResponse_access - The external ID of the server that the access is attached to.

Response Lenses

describeAccessResponse_httpStatus :: Lens' DescribeAccessResponse Int

The response's http status code.

describeAccessResponse_serverId :: Lens' DescribeAccessResponse Text

A system-assigned unique identifier for a server that has this access assigned.

describeAccessResponse_access :: Lens' DescribeAccessResponse DescribedAccess

The external ID of the server that the access is attached to.