Amazonka.Transfer.ImportSshPublicKey

Description

Adds a Secure Shell (SSH) public key to a user account identified by a UserName value assigned to the specific file transfer protocol-enabled server, identified by ServerId.

The response returns the UserName value, the ServerId value, and the name of the SshPublicKeyId.

Creating a Request

data ImportSshPublicKey

See: newImportSshPublicKey smart constructor.

Constructors

ImportSshPublicKey'

Fields

serverId :: Text

A system-assigned unique identifier for a server.

sshPublicKeyBody :: Text

The public key portion of an SSH key pair.

userName :: Text

The name of the user account that is assigned to one or more servers.

newImportSshPublicKey

Arguments

:: Text
$sel:serverId:ImportSshPublicKey'

-> Text
$sel:sshPublicKeyBody:ImportSshPublicKey'

-> Text
$sel:userName:ImportSshPublicKey'

-> ImportSshPublicKey

Create a value of ImportSshPublicKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverId:ImportSshPublicKey', importSshPublicKey_serverId - A system-assigned unique identifier for a server.

$sel:sshPublicKeyBody:ImportSshPublicKey', importSshPublicKey_sshPublicKeyBody - The public key portion of an SSH key pair.

$sel:userName:ImportSshPublicKey', importSshPublicKey_userName - The name of the user account that is assigned to one or more servers.

Request Lenses

importSshPublicKey_serverId :: Lens' ImportSshPublicKey Text

A system-assigned unique identifier for a server.

importSshPublicKey_sshPublicKeyBody :: Lens' ImportSshPublicKey Text

The public key portion of an SSH key pair.

importSshPublicKey_userName :: Lens' ImportSshPublicKey Text

The name of the user account that is assigned to one or more servers.

Destructuring the Response

data ImportSshPublicKeyResponse

Identifies the user, the server they belong to, and the identifier of the SSH public key associated with that user. A user can have more than one key on each server that they are associated with.

See: newImportSshPublicKeyResponse smart constructor.

Constructors

ImportSshPublicKeyResponse'

Fields

httpStatus :: Int

The response's http status code.

serverId :: Text

A system-assigned unique identifier for a server.

sshPublicKeyId :: Text

The name given to a public key by the system that was imported.

userName :: Text

A user name assigned to the ServerID value that you specified.

newImportSshPublicKeyResponse

Arguments

:: Int
$sel:httpStatus:ImportSshPublicKeyResponse'

-> Text
$sel:serverId:ImportSshPublicKey'

-> Text
$sel:sshPublicKeyId:ImportSshPublicKeyResponse'

-> Text
$sel:userName:ImportSshPublicKey'

-> ImportSshPublicKeyResponse

Create a value of ImportSshPublicKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ImportSshPublicKeyResponse', importSshPublicKeyResponse_httpStatus - The response's http status code.

$sel:serverId:ImportSshPublicKey', importSshPublicKeyResponse_serverId - A system-assigned unique identifier for a server.

$sel:sshPublicKeyId:ImportSshPublicKeyResponse', importSshPublicKeyResponse_sshPublicKeyId - The name given to a public key by the system that was imported.

$sel:userName:ImportSshPublicKey', importSshPublicKeyResponse_userName - A user name assigned to the ServerID value that you specified.

Response Lenses

importSshPublicKeyResponse_httpStatus :: Lens' ImportSshPublicKeyResponse Int

The response's http status code.

importSshPublicKeyResponse_serverId :: Lens' ImportSshPublicKeyResponse Text

A system-assigned unique identifier for a server.

importSshPublicKeyResponse_sshPublicKeyId :: Lens' ImportSshPublicKeyResponse Text

The name given to a public key by the system that was imported.

importSshPublicKeyResponse_userName :: Lens' ImportSshPublicKeyResponse Text

A user name assigned to the ServerID value that you specified.