Amazonka.Transfer.Types.DescribedWorkflow

Description

Documentation

data DescribedWorkflow

Describes the properties of the specified workflow

See: newDescribedWorkflow smart constructor.

Constructors

DescribedWorkflow'

Fields

onExceptionSteps :: Maybe [WorkflowStep]

Specifies the steps (actions) to take if errors are encountered during execution of the workflow.

steps :: Maybe [WorkflowStep]

Specifies the details for the steps that are in the specified workflow.

workflowId :: Maybe Text

A unique identifier for the workflow.

description :: Maybe Text

Specifies the text description for the workflow.

tags :: Maybe (NonEmpty Tag)

Key-value pairs that can be used to group and search for workflows. Tags are metadata attached to workflows for any purpose.

arn :: Text

Specifies the unique Amazon Resource Name (ARN) for the workflow.

newDescribedWorkflow

Arguments

:: Text
$sel:arn:DescribedWorkflow'

-> DescribedWorkflow

Create a value of DescribedWorkflow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:onExceptionSteps:DescribedWorkflow', describedWorkflow_onExceptionSteps - Specifies the steps (actions) to take if errors are encountered during execution of the workflow.

$sel:steps:DescribedWorkflow', describedWorkflow_steps - Specifies the details for the steps that are in the specified workflow.

$sel:workflowId:DescribedWorkflow', describedWorkflow_workflowId - A unique identifier for the workflow.

$sel:description:DescribedWorkflow', describedWorkflow_description - Specifies the text description for the workflow.

$sel:tags:DescribedWorkflow', describedWorkflow_tags - Key-value pairs that can be used to group and search for workflows. Tags are metadata attached to workflows for any purpose.

$sel:arn:DescribedWorkflow', describedWorkflow_arn - Specifies the unique Amazon Resource Name (ARN) for the workflow.

describedWorkflow_onExceptionSteps :: Lens' DescribedWorkflow (Maybe [WorkflowStep])

Specifies the steps (actions) to take if errors are encountered during execution of the workflow.

describedWorkflow_steps :: Lens' DescribedWorkflow (Maybe [WorkflowStep])

Specifies the details for the steps that are in the specified workflow.

describedWorkflow_workflowId :: Lens' DescribedWorkflow (Maybe Text)

A unique identifier for the workflow.

describedWorkflow_description :: Lens' DescribedWorkflow (Maybe Text)

Specifies the text description for the workflow.

describedWorkflow_tags :: Lens' DescribedWorkflow (Maybe (NonEmpty Tag))

Key-value pairs that can be used to group and search for workflows. Tags are metadata attached to workflows for any purpose.

describedWorkflow_arn :: Lens' DescribedWorkflow Text

Specifies the unique Amazon Resource Name (ARN) for the workflow.