Amazonka.Transfer.Types.EndpointDetails

Description

Documentation

data EndpointDetails

The virtual private cloud (VPC) endpoint settings that are configured for your file transfer protocol-enabled server. With a VPC endpoint, you can restrict access to your server and resources only within your VPC. To control incoming internet traffic, invoke the UpdateServer API and attach an Elastic IP address to your server's endpoint.

After May 19, 2021, you won't be able to create a server using EndpointType=VPC_ENDPOINT in your Amazon Web Servicesaccount if your account hasn't already done so before May 19, 2021. If you have already created servers with EndpointType=VPC_ENDPOINT in your Amazon Web Servicesaccount on or before May 19, 2021, you will not be affected. After this date, use EndpointType=VPC.

For more information, see https://docs.aws.amazon.com/transfer/latest/userguide/create-server-in-vpc.html#deprecate-vpc-endpoint.

See: newEndpointDetails smart constructor.

Constructors

EndpointDetails'

Fields

securityGroupIds :: Maybe [Text]

A list of security groups IDs that are available to attach to your server's endpoint.

This property can only be set when EndpointType is set to VPC.

You can edit the SecurityGroupIds property in the UpdateServer API only if you are changing the EndpointType from PUBLIC or VPC_ENDPOINT to VPC. To change security groups associated with your server's VPC endpoint after creation, use the Amazon EC2 ModifyVpcEndpoint API.

subnetIds :: Maybe [Text]

A list of subnet IDs that are required to host your server endpoint in your VPC.

This property can only be set when EndpointType is set to VPC.

vpcId :: Maybe Text

The VPC ID of the VPC in which a server's endpoint will be hosted.

This property can only be set when EndpointType is set to VPC.

addressAllocationIds :: Maybe [Text]

A list of address allocation IDs that are required to attach an Elastic IP address to your server's endpoint.

This property can only be set when EndpointType is set to VPC and it is only valid in the UpdateServer API.

vpcEndpointId :: Maybe Text

The ID of the VPC endpoint.

This property can only be set when EndpointType is set to VPC_ENDPOINT.

For more information, see https://docs.aws.amazon.com/transfer/latest/userguide/create-server-in-vpc.html#deprecate-vpc-endpoint.

newEndpointDetails :: EndpointDetails

Create a value of EndpointDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:EndpointDetails', endpointDetails_securityGroupIds - A list of security groups IDs that are available to attach to your server's endpoint.

This property can only be set when EndpointType is set to VPC.

You can edit the SecurityGroupIds property in the UpdateServer API only if you are changing the EndpointType from PUBLIC or VPC_ENDPOINT to VPC. To change security groups associated with your server's VPC endpoint after creation, use the Amazon EC2 ModifyVpcEndpoint API.

$sel:subnetIds:EndpointDetails', endpointDetails_subnetIds - A list of subnet IDs that are required to host your server endpoint in your VPC.

This property can only be set when EndpointType is set to VPC.

$sel:vpcId:EndpointDetails', endpointDetails_vpcId - The VPC ID of the VPC in which a server's endpoint will be hosted.

This property can only be set when EndpointType is set to VPC.

$sel:addressAllocationIds:EndpointDetails', endpointDetails_addressAllocationIds - A list of address allocation IDs that are required to attach an Elastic IP address to your server's endpoint.

This property can only be set when EndpointType is set to VPC and it is only valid in the UpdateServer API.

$sel:vpcEndpointId:EndpointDetails', endpointDetails_vpcEndpointId - The ID of the VPC endpoint.

This property can only be set when EndpointType is set to VPC_ENDPOINT.

For more information, see https://docs.aws.amazon.com/transfer/latest/userguide/create-server-in-vpc.html#deprecate-vpc-endpoint.

endpointDetails_securityGroupIds :: Lens' EndpointDetails (Maybe [Text])

A list of security groups IDs that are available to attach to your server's endpoint.

This property can only be set when EndpointType is set to VPC.

You can edit the SecurityGroupIds property in the UpdateServer API only if you are changing the EndpointType from PUBLIC or VPC_ENDPOINT to VPC. To change security groups associated with your server's VPC endpoint after creation, use the Amazon EC2 ModifyVpcEndpoint API.

endpointDetails_subnetIds :: Lens' EndpointDetails (Maybe [Text])

A list of subnet IDs that are required to host your server endpoint in your VPC.

This property can only be set when EndpointType is set to VPC.

endpointDetails_vpcId :: Lens' EndpointDetails (Maybe Text)

The VPC ID of the VPC in which a server's endpoint will be hosted.

This property can only be set when EndpointType is set to VPC.

endpointDetails_addressAllocationIds :: Lens' EndpointDetails (Maybe [Text])

A list of address allocation IDs that are required to attach an Elastic IP address to your server's endpoint.

This property can only be set when EndpointType is set to VPC and it is only valid in the UpdateServer API.

endpointDetails_vpcEndpointId :: Lens' EndpointDetails (Maybe Text)

The ID of the VPC endpoint.

This property can only be set when EndpointType is set to VPC_ENDPOINT.

For more information, see https://docs.aws.amazon.com/transfer/latest/userguide/create-server-in-vpc.html#deprecate-vpc-endpoint.