Amazonka.Transfer.Types.IdentityProviderDetails

Description

Documentation

data IdentityProviderDetails

Returns information related to the type of user authentication that is in use for a file transfer protocol-enabled server's users. A server can have only one method of authentication.

See: newIdentityProviderDetails smart constructor.

Constructors

IdentityProviderDetails'

Fields

invocationRole :: Maybe Text

Provides the type of InvocationRole used to authenticate the user account.

directoryId :: Maybe Text

The identifier of the Amazon Web ServicesDirectory Service directory that you want to stop sharing.

url :: Maybe Text

Provides the location of the service endpoint used to authenticate users.

newIdentityProviderDetails :: IdentityProviderDetails

Create a value of IdentityProviderDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invocationRole:IdentityProviderDetails', identityProviderDetails_invocationRole - Provides the type of InvocationRole used to authenticate the user account.

$sel:directoryId:IdentityProviderDetails', identityProviderDetails_directoryId - The identifier of the Amazon Web ServicesDirectory Service directory that you want to stop sharing.

$sel:url:IdentityProviderDetails', identityProviderDetails_url - Provides the location of the service endpoint used to authenticate users.

identityProviderDetails_invocationRole :: Lens' IdentityProviderDetails (Maybe Text)

Provides the type of InvocationRole used to authenticate the user account.

identityProviderDetails_directoryId :: Lens' IdentityProviderDetails (Maybe Text)

The identifier of the Amazon Web ServicesDirectory Service directory that you want to stop sharing.

identityProviderDetails_url :: Lens' IdentityProviderDetails (Maybe Text)

Provides the location of the service endpoint used to authenticate users.