Amazonka.Transfer.Types.InputFileLocation

Description

Documentation

data InputFileLocation

Specifies the location for the file being copied. Only applicable for the Copy type of workflow steps.

See: newInputFileLocation smart constructor.

Constructors

InputFileLocation'

Fields

efsFileLocation :: Maybe EfsFileLocation

Reserved for future use.

s3FileLocation :: Maybe S3InputFileLocation

Specifies the details for the S3 file being copied.

newInputFileLocation :: InputFileLocation

Create a value of InputFileLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:efsFileLocation:InputFileLocation', inputFileLocation_efsFileLocation - Reserved for future use.

$sel:s3FileLocation:InputFileLocation', inputFileLocation_s3FileLocation - Specifies the details for the S3 file being copied.

inputFileLocation_efsFileLocation :: Lens' InputFileLocation (Maybe EfsFileLocation)

Reserved for future use.

inputFileLocation_s3FileLocation :: Lens' InputFileLocation (Maybe S3InputFileLocation)

Specifies the details for the S3 file being copied.