Amazonka.Transfer.Types.ListedUser

Description

Documentation

data ListedUser

Returns properties of the user that you specify.

See: newListedUser smart constructor.

Constructors

ListedUser'

Fields

homeDirectoryType :: Maybe HomeDirectoryType

The type of landing directory (folder) you want your users' home directory to be when they log into the server. If you set it to PATH, the user will see the absolute Amazon S3 bucket or EFS paths as is in their file transfer protocol clients. If you set it LOGICAL, you need to provide mappings in the HomeDirectoryMappings for how you want to make Amazon S3 or EFS paths visible to your users.

userName :: Maybe Text

Specifies the name of the user whose ARN was specified. User names are used for authentication purposes.

role' :: Maybe Text

Specifies the Amazon Resource Name (ARN) of the IAM role that controls your users' access to your Amazon S3 bucket or EFS file system. The policies attached to this role determine the level of access that you want to provide your users when transferring files into and out of your Amazon S3 bucket or EFS file system. The IAM role should also contain a trust relationship that allows the server to access your resources when servicing your users' transfer requests.

The IAM role that controls your users' access to your Amazon S3 bucket for servers with Domain=S3, or your EFS file system for servers with Domain=EFS.

The policies attached to this role determine the level of access you want to provide your users when transferring files into and out of your S3 buckets or EFS file systems.

sshPublicKeyCount :: Maybe Int

Specifies the number of SSH public keys stored for the user you specified.

homeDirectory :: Maybe Text

The landing directory (folder) for a user when they log in to the server using the client.

A HomeDirectory example is /bucket_name/home/mydirectory.

arn :: Text

Provides the unique Amazon Resource Name (ARN) for the user that you want to learn about.

newListedUser

Arguments

:: Text
$sel:arn:ListedUser'

-> ListedUser

Create a value of ListedUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:homeDirectoryType:ListedUser', listedUser_homeDirectoryType - The type of landing directory (folder) you want your users' home directory to be when they log into the server. If you set it to PATH, the user will see the absolute Amazon S3 bucket or EFS paths as is in their file transfer protocol clients. If you set it LOGICAL, you need to provide mappings in the HomeDirectoryMappings for how you want to make Amazon S3 or EFS paths visible to your users.

$sel:userName:ListedUser', listedUser_userName - Specifies the name of the user whose ARN was specified. User names are used for authentication purposes.

$sel:role':ListedUser', listedUser_role - Specifies the Amazon Resource Name (ARN) of the IAM role that controls your users' access to your Amazon S3 bucket or EFS file system. The policies attached to this role determine the level of access that you want to provide your users when transferring files into and out of your Amazon S3 bucket or EFS file system. The IAM role should also contain a trust relationship that allows the server to access your resources when servicing your users' transfer requests.

The IAM role that controls your users' access to your Amazon S3 bucket for servers with Domain=S3, or your EFS file system for servers with Domain=EFS.

The policies attached to this role determine the level of access you want to provide your users when transferring files into and out of your S3 buckets or EFS file systems.

$sel:sshPublicKeyCount:ListedUser', listedUser_sshPublicKeyCount - Specifies the number of SSH public keys stored for the user you specified.

$sel:homeDirectory:ListedUser', listedUser_homeDirectory - The landing directory (folder) for a user when they log in to the server using the client.

A HomeDirectory example is /bucket_name/home/mydirectory.

$sel:arn:ListedUser', listedUser_arn - Provides the unique Amazon Resource Name (ARN) for the user that you want to learn about.

listedUser_homeDirectoryType :: Lens' ListedUser (Maybe HomeDirectoryType)

The type of landing directory (folder) you want your users' home directory to be when they log into the server. If you set it to PATH, the user will see the absolute Amazon S3 bucket or EFS paths as is in their file transfer protocol clients. If you set it LOGICAL, you need to provide mappings in the HomeDirectoryMappings for how you want to make Amazon S3 or EFS paths visible to your users.

listedUser_userName :: Lens' ListedUser (Maybe Text)

Specifies the name of the user whose ARN was specified. User names are used for authentication purposes.

listedUser_role :: Lens' ListedUser (Maybe Text)

Specifies the Amazon Resource Name (ARN) of the IAM role that controls your users' access to your Amazon S3 bucket or EFS file system. The policies attached to this role determine the level of access that you want to provide your users when transferring files into and out of your Amazon S3 bucket or EFS file system. The IAM role should also contain a trust relationship that allows the server to access your resources when servicing your users' transfer requests.

The IAM role that controls your users' access to your Amazon S3 bucket for servers with Domain=S3, or your EFS file system for servers with Domain=EFS.

The policies attached to this role determine the level of access you want to provide your users when transferring files into and out of your S3 buckets or EFS file systems.

listedUser_sshPublicKeyCount :: Lens' ListedUser (Maybe Int)

Specifies the number of SSH public keys stored for the user you specified.

listedUser_homeDirectory :: Lens' ListedUser (Maybe Text)

The landing directory (folder) for a user when they log in to the server using the client.

A HomeDirectory example is /bucket_name/home/mydirectory.

listedUser_arn :: Lens' ListedUser Text

Provides the unique Amazon Resource Name (ARN) for the user that you want to learn about.