Amazonka.Transfer.Types.PosixProfile

Description

Documentation

data PosixProfile

The full POSIX identity, including user ID (Uid), group ID (Gid), and any secondary groups IDs (SecondaryGids), that controls your users' access to your Amazon EFS file systems. The POSIX permissions that are set on files and directories in your file system determine the level of access your users get when transferring files into and out of your Amazon EFS file systems.

See: newPosixProfile smart constructor.

Constructors

PosixProfile'

Fields

secondaryGids :: Maybe [Natural]

The secondary POSIX group IDs used for all EFS operations by this user.

uid :: Natural

The POSIX user ID used for all EFS operations by this user.

gid :: Natural

The POSIX group ID used for all EFS operations by this user.

newPosixProfile

Arguments

:: Natural
$sel:uid:PosixProfile'

-> Natural
$sel:gid:PosixProfile'

-> PosixProfile

Create a value of PosixProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secondaryGids:PosixProfile', posixProfile_secondaryGids - The secondary POSIX group IDs used for all EFS operations by this user.

$sel:uid:PosixProfile', posixProfile_uid - The POSIX user ID used for all EFS operations by this user.

$sel:gid:PosixProfile', posixProfile_gid - The POSIX group ID used for all EFS operations by this user.

posixProfile_secondaryGids :: Lens' PosixProfile (Maybe [Natural])

The secondary POSIX group IDs used for all EFS operations by this user.

posixProfile_uid :: Lens' PosixProfile Natural

The POSIX user ID used for all EFS operations by this user.

posixProfile_gid :: Lens' PosixProfile Natural

The POSIX group ID used for all EFS operations by this user.