Amazonka.Translate.StartTextTranslationJob

Description

Starts an asynchronous batch translation job. Batch translation jobs can be used to translate large volumes of text across multiple documents at once. For more information, see async.

Batch translation jobs can be described with the DescribeTextTranslationJob operation, listed with the ListTextTranslationJobs operation, and stopped with the StopTextTranslationJob operation.

Amazon Translate does not support batch translation of multiple source languages at once.

Creating a Request

data StartTextTranslationJob

See: newStartTextTranslationJob smart constructor.

Constructors

StartTextTranslationJob'

Fields

jobName :: Maybe Text

The name of the batch translation job to be performed.

parallelDataNames :: Maybe [Text]

The names of the parallel data resources to use in the batch translation job. For a list of available parallel data resources, use the ListParallelData operation.

terminologyNames :: Maybe [Text]

The name of the terminology to use in the batch translation job. For a list of available terminologies, use the ListTerminologies operation.

inputDataConfig :: InputDataConfig

Specifies the format and S3 location of the input documents for the translation job.

outputDataConfig :: OutputDataConfig

Specifies the S3 folder to which your job output will be saved.

dataAccessRoleArn :: Text

The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that grants Amazon Translate read access to your input data. For more nformation, see identity-and-access-management.

sourceLanguageCode :: Text

The language code of the input language. For a list of language codes, see what-is-languages.

Amazon Translate does not automatically detect a source language during batch translation jobs.

targetLanguageCodes :: NonEmpty Text

The language code of the output language.

clientToken :: Text

A unique identifier for the request. This token is auto-generated when using the Amazon Translate SDK.

newStartTextTranslationJob

Arguments

:: InputDataConfig
$sel:inputDataConfig:StartTextTranslationJob'

-> OutputDataConfig
$sel:outputDataConfig:StartTextTranslationJob'

-> Text
$sel:dataAccessRoleArn:StartTextTranslationJob'

-> Text
$sel:sourceLanguageCode:StartTextTranslationJob'

-> NonEmpty Text
$sel:targetLanguageCodes:StartTextTranslationJob'

-> Text
$sel:clientToken:StartTextTranslationJob'

-> StartTextTranslationJob

Create a value of StartTextTranslationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobName:StartTextTranslationJob', startTextTranslationJob_jobName - The name of the batch translation job to be performed.

$sel:parallelDataNames:StartTextTranslationJob', startTextTranslationJob_parallelDataNames - The names of the parallel data resources to use in the batch translation job. For a list of available parallel data resources, use the ListParallelData operation.

$sel:terminologyNames:StartTextTranslationJob', startTextTranslationJob_terminologyNames - The name of the terminology to use in the batch translation job. For a list of available terminologies, use the ListTerminologies operation.

$sel:inputDataConfig:StartTextTranslationJob', startTextTranslationJob_inputDataConfig - Specifies the format and S3 location of the input documents for the translation job.

$sel:outputDataConfig:StartTextTranslationJob', startTextTranslationJob_outputDataConfig - Specifies the S3 folder to which your job output will be saved.

$sel:dataAccessRoleArn:StartTextTranslationJob', startTextTranslationJob_dataAccessRoleArn - The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that grants Amazon Translate read access to your input data. For more nformation, see identity-and-access-management.

$sel:sourceLanguageCode:StartTextTranslationJob', startTextTranslationJob_sourceLanguageCode - The language code of the input language. For a list of language codes, see what-is-languages.

Amazon Translate does not automatically detect a source language during batch translation jobs.

$sel:targetLanguageCodes:StartTextTranslationJob', startTextTranslationJob_targetLanguageCodes - The language code of the output language.

$sel:clientToken:StartTextTranslationJob', startTextTranslationJob_clientToken - A unique identifier for the request. This token is auto-generated when using the Amazon Translate SDK.

Request Lenses

startTextTranslationJob_jobName :: Lens' StartTextTranslationJob (Maybe Text)

The name of the batch translation job to be performed.

startTextTranslationJob_parallelDataNames :: Lens' StartTextTranslationJob (Maybe [Text])

The names of the parallel data resources to use in the batch translation job. For a list of available parallel data resources, use the ListParallelData operation.

startTextTranslationJob_terminologyNames :: Lens' StartTextTranslationJob (Maybe [Text])

The name of the terminology to use in the batch translation job. For a list of available terminologies, use the ListTerminologies operation.

startTextTranslationJob_inputDataConfig :: Lens' StartTextTranslationJob InputDataConfig

Specifies the format and S3 location of the input documents for the translation job.

startTextTranslationJob_outputDataConfig :: Lens' StartTextTranslationJob OutputDataConfig

Specifies the S3 folder to which your job output will be saved.

startTextTranslationJob_dataAccessRoleArn :: Lens' StartTextTranslationJob Text

The Amazon Resource Name (ARN) of an AWS Identity Access and Management (IAM) role that grants Amazon Translate read access to your input data. For more nformation, see identity-and-access-management.

startTextTranslationJob_sourceLanguageCode :: Lens' StartTextTranslationJob Text

The language code of the input language. For a list of language codes, see what-is-languages.

Amazon Translate does not automatically detect a source language during batch translation jobs.

startTextTranslationJob_targetLanguageCodes :: Lens' StartTextTranslationJob (NonEmpty Text)

The language code of the output language.

startTextTranslationJob_clientToken :: Lens' StartTextTranslationJob Text

A unique identifier for the request. This token is auto-generated when using the Amazon Translate SDK.

Destructuring the Response

data StartTextTranslationJobResponse

See: newStartTextTranslationJobResponse smart constructor.

Constructors

StartTextTranslationJobResponse'

Fields

jobId :: Maybe Text

The identifier generated for the job. To get the status of a job, use this ID with the DescribeTextTranslationJob operation.

jobStatus :: Maybe JobStatus

The status of the job. Possible values include:

SUBMITTED - The job has been received and is queued for processing.

IN_PROGRESS - Amazon Translate is processing the job.

COMPLETED - The job was successfully completed and the output is available.

COMPLETED_WITH_ERROR - The job was completed with errors. The errors can be analyzed in the job's output.

FAILED - The job did not complete. To get details, use the DescribeTextTranslationJob operation.

STOP_REQUESTED - The user who started the job has requested that it be stopped.

STOPPED - The job has been stopped.

httpStatus :: Int

The response's http status code.

newStartTextTranslationJobResponse

Arguments

:: Int
$sel:httpStatus:StartTextTranslationJobResponse'

-> StartTextTranslationJobResponse

Create a value of StartTextTranslationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:StartTextTranslationJobResponse', startTextTranslationJobResponse_jobId - The identifier generated for the job. To get the status of a job, use this ID with the DescribeTextTranslationJob operation.

$sel:jobStatus:StartTextTranslationJobResponse', startTextTranslationJobResponse_jobStatus - The status of the job. Possible values include:

SUBMITTED - The job has been received and is queued for processing.

IN_PROGRESS - Amazon Translate is processing the job.

COMPLETED - The job was successfully completed and the output is available.

COMPLETED_WITH_ERROR - The job was completed with errors. The errors can be analyzed in the job's output.

FAILED - The job did not complete. To get details, use the DescribeTextTranslationJob operation.

STOP_REQUESTED - The user who started the job has requested that it be stopped.

STOPPED - The job has been stopped.

$sel:httpStatus:StartTextTranslationJobResponse', startTextTranslationJobResponse_httpStatus - The response's http status code.

Response Lenses

startTextTranslationJobResponse_jobId :: Lens' StartTextTranslationJobResponse (Maybe Text)

The identifier generated for the job. To get the status of a job, use this ID with the DescribeTextTranslationJob operation.

startTextTranslationJobResponse_jobStatus :: Lens' StartTextTranslationJobResponse (Maybe JobStatus)

The status of the job. Possible values include:

SUBMITTED - The job has been received and is queued for processing.

IN_PROGRESS - Amazon Translate is processing the job.

COMPLETED - The job was successfully completed and the output is available.

COMPLETED_WITH_ERROR - The job was completed with errors. The errors can be analyzed in the job's output.

FAILED - The job did not complete. To get details, use the DescribeTextTranslationJob operation.

STOP_REQUESTED - The user who started the job has requested that it be stopped.

STOPPED - The job has been stopped.

startTextTranslationJobResponse_httpStatus :: Lens' StartTextTranslationJobResponse Int

The response's http status code.