Amazonka.Translate.Types.ParallelDataProperties

Description

Documentation

data ParallelDataProperties

The properties of a parallel data resource.

See: newParallelDataProperties smart constructor.

Constructors

ParallelDataProperties'

Fields

status :: Maybe ParallelDataStatus

The status of the parallel data resource. When the parallel data is ready for you to use, the status is ACTIVE.

lastUpdatedAt :: Maybe POSIX

The time at which the parallel data resource was last updated.

importedRecordCount :: Maybe Integer

The number of records successfully imported from the parallel data input file.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the parallel data resource.

targetLanguageCodes :: Maybe [Text]

The language codes for the target languages available in the parallel data file. All possible target languages are returned as an array.

createdAt :: Maybe POSIX

The time at which the parallel data resource was created.

failedRecordCount :: Maybe Integer

The number of records unsuccessfully imported from the parallel data input file.

importedDataSize :: Maybe Integer

The number of UTF-8 characters that Amazon Translate imported from the parallel data input file. This number includes only the characters in your translation examples. It does not include characters that are used to format your file. For example, if you provided a Translation Memory Exchange (.tmx) file, this number does not include the tags.

name :: Maybe Text

The custom name assigned to the parallel data resource.

sourceLanguageCode :: Maybe Text

The source language of the translations in the parallel data file.

latestUpdateAttemptAt :: Maybe POSIX

The time that the most recent update was attempted.

encryptionKey :: Maybe EncryptionKey

latestUpdateAttemptStatus :: Maybe ParallelDataStatus

The status of the most recent update attempt for the parallel data resource.

message :: Maybe Text

Additional information from Amazon Translate about the parallel data resource.

description :: Maybe Text

The description assigned to the parallel data resource.

skippedRecordCount :: Maybe Integer

The number of items in the input file that Amazon Translate skipped when you created or updated the parallel data resource. For example, Amazon Translate skips empty records, empty target texts, and empty lines.

parallelDataConfig :: Maybe ParallelDataConfig

Specifies the format and S3 location of the parallel data input file.

newParallelDataProperties :: ParallelDataProperties

Create a value of ParallelDataProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ParallelDataProperties', parallelDataProperties_status - The status of the parallel data resource. When the parallel data is ready for you to use, the status is ACTIVE.

$sel:lastUpdatedAt:ParallelDataProperties', parallelDataProperties_lastUpdatedAt - The time at which the parallel data resource was last updated.

$sel:importedRecordCount:ParallelDataProperties', parallelDataProperties_importedRecordCount - The number of records successfully imported from the parallel data input file.

$sel:arn:ParallelDataProperties', parallelDataProperties_arn - The Amazon Resource Name (ARN) of the parallel data resource.

$sel:targetLanguageCodes:ParallelDataProperties', parallelDataProperties_targetLanguageCodes - The language codes for the target languages available in the parallel data file. All possible target languages are returned as an array.

$sel:createdAt:ParallelDataProperties', parallelDataProperties_createdAt - The time at which the parallel data resource was created.

$sel:failedRecordCount:ParallelDataProperties', parallelDataProperties_failedRecordCount - The number of records unsuccessfully imported from the parallel data input file.

$sel:importedDataSize:ParallelDataProperties', parallelDataProperties_importedDataSize - The number of UTF-8 characters that Amazon Translate imported from the parallel data input file. This number includes only the characters in your translation examples. It does not include characters that are used to format your file. For example, if you provided a Translation Memory Exchange (.tmx) file, this number does not include the tags.

$sel:name:ParallelDataProperties', parallelDataProperties_name - The custom name assigned to the parallel data resource.

$sel:sourceLanguageCode:ParallelDataProperties', parallelDataProperties_sourceLanguageCode - The source language of the translations in the parallel data file.

$sel:latestUpdateAttemptAt:ParallelDataProperties', parallelDataProperties_latestUpdateAttemptAt - The time that the most recent update was attempted.

$sel:encryptionKey:ParallelDataProperties', parallelDataProperties_encryptionKey - Undocumented member.

$sel:latestUpdateAttemptStatus:ParallelDataProperties', parallelDataProperties_latestUpdateAttemptStatus - The status of the most recent update attempt for the parallel data resource.

$sel:message:ParallelDataProperties', parallelDataProperties_message - Additional information from Amazon Translate about the parallel data resource.

$sel:description:ParallelDataProperties', parallelDataProperties_description - The description assigned to the parallel data resource.

$sel:skippedRecordCount:ParallelDataProperties', parallelDataProperties_skippedRecordCount - The number of items in the input file that Amazon Translate skipped when you created or updated the parallel data resource. For example, Amazon Translate skips empty records, empty target texts, and empty lines.

$sel:parallelDataConfig:ParallelDataProperties', parallelDataProperties_parallelDataConfig - Specifies the format and S3 location of the parallel data input file.

parallelDataProperties_status :: Lens' ParallelDataProperties (Maybe ParallelDataStatus)

The status of the parallel data resource. When the parallel data is ready for you to use, the status is ACTIVE.

parallelDataProperties_lastUpdatedAt :: Lens' ParallelDataProperties (Maybe UTCTime)

The time at which the parallel data resource was last updated.

parallelDataProperties_importedRecordCount :: Lens' ParallelDataProperties (Maybe Integer)

The number of records successfully imported from the parallel data input file.

parallelDataProperties_arn :: Lens' ParallelDataProperties (Maybe Text)

The Amazon Resource Name (ARN) of the parallel data resource.

parallelDataProperties_targetLanguageCodes :: Lens' ParallelDataProperties (Maybe [Text])

The language codes for the target languages available in the parallel data file. All possible target languages are returned as an array.

parallelDataProperties_createdAt :: Lens' ParallelDataProperties (Maybe UTCTime)

The time at which the parallel data resource was created.

parallelDataProperties_failedRecordCount :: Lens' ParallelDataProperties (Maybe Integer)

The number of records unsuccessfully imported from the parallel data input file.

parallelDataProperties_importedDataSize :: Lens' ParallelDataProperties (Maybe Integer)

The number of UTF-8 characters that Amazon Translate imported from the parallel data input file. This number includes only the characters in your translation examples. It does not include characters that are used to format your file. For example, if you provided a Translation Memory Exchange (.tmx) file, this number does not include the tags.

parallelDataProperties_name :: Lens' ParallelDataProperties (Maybe Text)

The custom name assigned to the parallel data resource.

parallelDataProperties_sourceLanguageCode :: Lens' ParallelDataProperties (Maybe Text)

The source language of the translations in the parallel data file.

parallelDataProperties_latestUpdateAttemptAt :: Lens' ParallelDataProperties (Maybe UTCTime)

The time that the most recent update was attempted.

parallelDataProperties_encryptionKey :: Lens' ParallelDataProperties (Maybe EncryptionKey)

Undocumented member.

parallelDataProperties_latestUpdateAttemptStatus :: Lens' ParallelDataProperties (Maybe ParallelDataStatus)

The status of the most recent update attempt for the parallel data resource.

parallelDataProperties_message :: Lens' ParallelDataProperties (Maybe Text)

Additional information from Amazon Translate about the parallel data resource.

parallelDataProperties_description :: Lens' ParallelDataProperties (Maybe Text)

The description assigned to the parallel data resource.

parallelDataProperties_skippedRecordCount :: Lens' ParallelDataProperties (Maybe Integer)

The number of items in the input file that Amazon Translate skipped when you created or updated the parallel data resource. For example, Amazon Translate skips empty records, empty target texts, and empty lines.

parallelDataProperties_parallelDataConfig :: Lens' ParallelDataProperties (Maybe ParallelDataConfig)

Specifies the format and S3 location of the parallel data input file.