Amazonka.Translate.Types.TextTranslationJobFilter

Description

Documentation

data TextTranslationJobFilter

Provides information for filtering a list of translation jobs. For more information, see ListTextTranslationJobs.

See: newTextTranslationJobFilter smart constructor.

Constructors

TextTranslationJobFilter'

Fields

submittedBeforeTime :: Maybe POSIX

Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

submittedAfterTime :: Maybe POSIX

Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

jobName :: Maybe Text

Filters the list of jobs by name.

jobStatus :: Maybe JobStatus

Filters the list of jobs based by job status.

newTextTranslationJobFilter :: TextTranslationJobFilter

Create a value of TextTranslationJobFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:submittedBeforeTime:TextTranslationJobFilter', textTranslationJobFilter_submittedBeforeTime - Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

$sel:submittedAfterTime:TextTranslationJobFilter', textTranslationJobFilter_submittedAfterTime - Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

$sel:jobName:TextTranslationJobFilter', textTranslationJobFilter_jobName - Filters the list of jobs by name.

$sel:jobStatus:TextTranslationJobFilter', textTranslationJobFilter_jobStatus - Filters the list of jobs based by job status.

textTranslationJobFilter_submittedBeforeTime :: Lens' TextTranslationJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted before the specified time. Jobs are returned in ascending order, oldest to newest.

textTranslationJobFilter_submittedAfterTime :: Lens' TextTranslationJobFilter (Maybe UTCTime)

Filters the list of jobs based on the time that the job was submitted for processing and returns only the jobs submitted after the specified time. Jobs are returned in descending order, newest to oldest.

textTranslationJobFilter_jobName :: Lens' TextTranslationJobFilter (Maybe Text)

Filters the list of jobs by name.

textTranslationJobFilter_jobStatus :: Lens' TextTranslationJobFilter (Maybe JobStatus)

Filters the list of jobs based by job status.