Amazonka.VoiceId.CreateDomain

Description

Creates a domain that contains all Amazon Connect Voice ID data, such as speakers, fraudsters, customer audio, and voiceprints.

Creating a Request

data CreateDomain

See: newCreateDomain smart constructor.

Constructors

CreateDomain'

Fields

clientToken :: Maybe Text

The idempotency token for creating a new domain. If not provided, Amazon Web Services SDK populates this field.

description :: Maybe (Sensitive Text)

A brief description of this domain.

tags :: Maybe [Tag]

A list of tags you want added to the domain.

name :: Sensitive Text

The name of the domain.

serverSideEncryptionConfiguration :: ServerSideEncryptionConfiguration

The configuration, containing the KMS Key Identifier, to be used by Voice ID for the server-side encryption of your data. Refer to Amazon Connect VoiceID encryption at rest for more details on how the KMS Key is used.

newCreateDomain

Arguments

:: Text
$sel:name:CreateDomain'

-> ServerSideEncryptionConfiguration
$sel:serverSideEncryptionConfiguration:CreateDomain'

-> CreateDomain

Create a value of CreateDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateDomain', createDomain_clientToken - The idempotency token for creating a new domain. If not provided, Amazon Web Services SDK populates this field.

$sel:description:CreateDomain', createDomain_description - A brief description of this domain.

$sel:tags:CreateDomain', createDomain_tags - A list of tags you want added to the domain.

$sel:name:CreateDomain', createDomain_name - The name of the domain.

$sel:serverSideEncryptionConfiguration:CreateDomain', createDomain_serverSideEncryptionConfiguration - The configuration, containing the KMS Key Identifier, to be used by Voice ID for the server-side encryption of your data. Refer to Amazon Connect VoiceID encryption at rest for more details on how the KMS Key is used.

Request Lenses

createDomain_clientToken :: Lens' CreateDomain (Maybe Text)

The idempotency token for creating a new domain. If not provided, Amazon Web Services SDK populates this field.

createDomain_description :: Lens' CreateDomain (Maybe Text)

A brief description of this domain.

createDomain_tags :: Lens' CreateDomain (Maybe [Tag])

A list of tags you want added to the domain.

createDomain_name :: Lens' CreateDomain Text

The name of the domain.

createDomain_serverSideEncryptionConfiguration :: Lens' CreateDomain ServerSideEncryptionConfiguration

The configuration, containing the KMS Key Identifier, to be used by Voice ID for the server-side encryption of your data. Refer to Amazon Connect VoiceID encryption at rest for more details on how the KMS Key is used.

Destructuring the Response

data CreateDomainResponse

See: newCreateDomainResponse smart constructor.

Constructors

CreateDomainResponse'

Fields

domain :: Maybe Domain

Information about the newly created domain.

httpStatus :: Int

The response's http status code.

newCreateDomainResponse

Arguments

:: Int
$sel:httpStatus:CreateDomainResponse'

-> CreateDomainResponse

Create a value of CreateDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domain:CreateDomainResponse', createDomainResponse_domain - Information about the newly created domain.

$sel:httpStatus:CreateDomainResponse', createDomainResponse_httpStatus - The response's http status code.

Response Lenses

createDomainResponse_domain :: Lens' CreateDomainResponse (Maybe Domain)

Information about the newly created domain.

createDomainResponse_httpStatus :: Lens' CreateDomainResponse Int

The response's http status code.