Amazonka.VoiceId.Lens

Description

Operations

DescribeFraudsterRegistrationJob

describeFraudsterRegistrationJob_domainId :: Lens' DescribeFraudsterRegistrationJob Text

The identifier for the domain containing the fraudster registration job.

describeFraudsterRegistrationJob_jobId :: Lens' DescribeFraudsterRegistrationJob Text

The identifier for the fraudster registration job you are describing.

describeFraudsterRegistrationJobResponse_job :: Lens' DescribeFraudsterRegistrationJobResponse (Maybe FraudsterRegistrationJob)

Contains details about the specified fraudster registration job.

describeFraudsterRegistrationJobResponse_httpStatus :: Lens' DescribeFraudsterRegistrationJobResponse Int

The response's http status code.

DeleteSpeaker

deleteSpeaker_domainId :: Lens' DeleteSpeaker Text

The identifier of the domain containing the speaker.

deleteSpeaker_speakerId :: Lens' DeleteSpeaker Text

The identifier of the speaker you want to delete.

ListSpeakers

listSpeakers_nextToken :: Lens' ListSpeakers (Maybe Text)

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

listSpeakers_maxResults :: Lens' ListSpeakers (Maybe Natural)

The maximum number of results that are returned per call. You can use NextToken to obtain further pages of results. The default is 100; the maximum allowed page size is also 100.

listSpeakers_domainId :: Lens' ListSpeakers Text

The identifier of the domain.

listSpeakersResponse_speakerSummaries :: Lens' ListSpeakersResponse (Maybe [SpeakerSummary])

A list containing details about each speaker in the Amazon Web Services account.

listSpeakersResponse_nextToken :: Lens' ListSpeakersResponse (Maybe Text)

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

listSpeakersResponse_httpStatus :: Lens' ListSpeakersResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the Voice ID resource for which you want to list the tags.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

The list of tags associated with the specified resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

EvaluateSession

evaluateSession_domainId :: Lens' EvaluateSession Text

The identifier of the domain where the session started.

evaluateSession_sessionNameOrId :: Lens' EvaluateSession Text

The session identifier, or name of the session, that you want to evaluate. In Voice ID integration, this is the Contact-Id.

evaluateSessionResponse_fraudDetectionResult :: Lens' EvaluateSessionResponse (Maybe FraudDetectionResult)

Details resulting from the fraud detection process, such as fraud detection decision and risk score.

evaluateSessionResponse_streamingStatus :: Lens' EvaluateSessionResponse (Maybe StreamingStatus)

The current status of audio streaming for this session. This field is useful to infer next steps when the Authentication or Fraud Detection results are empty or the decision is NOT_ENOUGH_SPEECH. In this situation, if the StreamingStatus is ONGOING/PENDING_CONFIGURATION, it can mean that the client should call the API again later, once Voice ID has enough audio to produce a result. If the decision remains NOT_ENOUGH_SPEECH even after StreamingStatus is ENDED, it means that the previously streamed session did not have enough speech to perform evaluation, and a new streaming session is needed to try again.

evaluateSessionResponse_authenticationResult :: Lens' EvaluateSessionResponse (Maybe AuthenticationResult)

Details resulting from the authentication process, such as authentication decision and authentication score.

evaluateSessionResponse_domainId :: Lens' EvaluateSessionResponse (Maybe Text)

The identifier of the domain containing the session.

evaluateSessionResponse_sessionId :: Lens' EvaluateSessionResponse (Maybe Text)

The service-generated identifier of the session.

evaluateSessionResponse_sessionName :: Lens' EvaluateSessionResponse (Maybe Text)

The client-provided name of the session.

evaluateSessionResponse_httpStatus :: Lens' EvaluateSessionResponse Int

The response's http status code.

DescribeSpeakerEnrollmentJob

describeSpeakerEnrollmentJob_domainId :: Lens' DescribeSpeakerEnrollmentJob Text

The identifier of the domain containing the speaker enrollment job.

describeSpeakerEnrollmentJob_jobId :: Lens' DescribeSpeakerEnrollmentJob Text

The identifier of the speaker enrollment job you are describing.

describeSpeakerEnrollmentJobResponse_job :: Lens' DescribeSpeakerEnrollmentJobResponse (Maybe SpeakerEnrollmentJob)

Contains details about the specified speaker enrollment job.

describeSpeakerEnrollmentJobResponse_httpStatus :: Lens' DescribeSpeakerEnrollmentJobResponse Int

The response's http status code.

DeleteFraudster

deleteFraudster_domainId :: Lens' DeleteFraudster Text

The identifier of the domain containing the fraudster.

deleteFraudster_fraudsterId :: Lens' DeleteFraudster Text

The identifier of the fraudster you want to delete.

ListFraudsterRegistrationJobs

listFraudsterRegistrationJobs_nextToken :: Lens' ListFraudsterRegistrationJobs (Maybe Text)

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

listFraudsterRegistrationJobs_jobStatus :: Lens' ListFraudsterRegistrationJobs (Maybe FraudsterRegistrationJobStatus)

Provides the status of your fraudster registration job.

listFraudsterRegistrationJobs_maxResults :: Lens' ListFraudsterRegistrationJobs (Maybe Natural)

The maximum number of results that are returned per call. You can use NextToken to obtain further pages of results. The default is 100; the maximum allowed page size is also 100.

listFraudsterRegistrationJobs_domainId :: Lens' ListFraudsterRegistrationJobs Text

The identifier of the domain containing the fraudster registration Jobs.

listFraudsterRegistrationJobsResponse_nextToken :: Lens' ListFraudsterRegistrationJobsResponse (Maybe Text)

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

listFraudsterRegistrationJobsResponse_jobSummaries :: Lens' ListFraudsterRegistrationJobsResponse (Maybe [FraudsterRegistrationJobSummary])

A list containing details about each specified fraudster registration job.

listFraudsterRegistrationJobsResponse_httpStatus :: Lens' ListFraudsterRegistrationJobsResponse Int

The response's http status code.

DescribeFraudster

describeFraudster_domainId :: Lens' DescribeFraudster Text

The identifier of the domain containing the fraudster.

describeFraudster_fraudsterId :: Lens' DescribeFraudster Text

The identifier of the fraudster you are describing.

describeFraudsterResponse_fraudster :: Lens' DescribeFraudsterResponse (Maybe Fraudster)

Information about the specified fraudster.

describeFraudsterResponse_httpStatus :: Lens' DescribeFraudsterResponse Int

The response's http status code.

ListSpeakerEnrollmentJobs

listSpeakerEnrollmentJobs_nextToken :: Lens' ListSpeakerEnrollmentJobs (Maybe Text)

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

listSpeakerEnrollmentJobs_jobStatus :: Lens' ListSpeakerEnrollmentJobs (Maybe SpeakerEnrollmentJobStatus)

Provides the status of your speaker enrollment Job.

listSpeakerEnrollmentJobs_maxResults :: Lens' ListSpeakerEnrollmentJobs (Maybe Natural)

The maximum number of results that are returned per call. You can use NextToken to obtain further pages of results. The default is 100; the maximum allowed page size is also 100.

listSpeakerEnrollmentJobs_domainId :: Lens' ListSpeakerEnrollmentJobs Text

The identifier of the domain containing the speaker enrollment jobs.

listSpeakerEnrollmentJobsResponse_nextToken :: Lens' ListSpeakerEnrollmentJobsResponse (Maybe Text)

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

listSpeakerEnrollmentJobsResponse_jobSummaries :: Lens' ListSpeakerEnrollmentJobsResponse (Maybe [SpeakerEnrollmentJobSummary])

A list containing details about each specified speaker enrollment job.

listSpeakerEnrollmentJobsResponse_httpStatus :: Lens' ListSpeakerEnrollmentJobsResponse Int

The response's http status code.

CreateDomain

createDomain_clientToken :: Lens' CreateDomain (Maybe Text)

The idempotency token for creating a new domain. If not provided, Amazon Web Services SDK populates this field.

createDomain_description :: Lens' CreateDomain (Maybe Text)

A brief description of this domain.

createDomain_tags :: Lens' CreateDomain (Maybe [Tag])

A list of tags you want added to the domain.

createDomain_name :: Lens' CreateDomain Text

The name of the domain.

createDomain_serverSideEncryptionConfiguration :: Lens' CreateDomain ServerSideEncryptionConfiguration

The configuration, containing the KMS Key Identifier, to be used by Voice ID for the server-side encryption of your data. Refer to Amazon Connect VoiceID encryption at rest for more details on how the KMS Key is used.

createDomainResponse_domain :: Lens' CreateDomainResponse (Maybe Domain)

Information about the newly created domain.

createDomainResponse_httpStatus :: Lens' CreateDomainResponse Int

The response's http status code.

StartFraudsterRegistrationJob

startFraudsterRegistrationJob_clientToken :: Lens' StartFraudsterRegistrationJob (Maybe Text)

The idempotency token for starting a new fraudster registration job. If not provided, Amazon Web Services SDK populates this field.

startFraudsterRegistrationJob_jobName :: Lens' StartFraudsterRegistrationJob (Maybe Text)

The name of the new fraudster registration job.

startFraudsterRegistrationJob_registrationConfig :: Lens' StartFraudsterRegistrationJob (Maybe RegistrationConfig)

The registration config containing details such as the action to take when a duplicate fraudster is detected, and the similarity threshold to use for detecting a duplicate fraudster.

startFraudsterRegistrationJob_dataAccessRoleArn :: Lens' StartFraudsterRegistrationJob Text

The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the Job output file. Refer to the Create and edit a fraudster watchlist documentation for the permissions needed in this role.

startFraudsterRegistrationJob_domainId :: Lens' StartFraudsterRegistrationJob Text

The identifier of the domain containing the fraudster registration job and in which the fraudsters are registered.

startFraudsterRegistrationJob_inputDataConfig :: Lens' StartFraudsterRegistrationJob InputDataConfig

The input data config containing an S3 URI for the input manifest file that contains the list of fraudster registration requests.

startFraudsterRegistrationJob_outputDataConfig :: Lens' StartFraudsterRegistrationJob OutputDataConfig

The output data config containing the S3 location where Voice ID writes the job output file; you must also include a KMS Key ID to encrypt the file.

startFraudsterRegistrationJobResponse_job :: Lens' StartFraudsterRegistrationJobResponse (Maybe FraudsterRegistrationJob)

Details about the started fraudster registration job.

startFraudsterRegistrationJobResponse_httpStatus :: Lens' StartFraudsterRegistrationJobResponse Int

The response's http status code.

DescribeDomain

describeDomain_domainId :: Lens' DescribeDomain Text

The identifier of the domain you are describing.

describeDomainResponse_domain :: Lens' DescribeDomainResponse (Maybe Domain)

Information about the specified domain.

describeDomainResponse_httpStatus :: Lens' DescribeDomainResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the Voice ID resource you want to tag.

tagResource_tags :: Lens' TagResource [Tag]

The list of tags to assign to the specified resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

StartSpeakerEnrollmentJob

startSpeakerEnrollmentJob_clientToken :: Lens' StartSpeakerEnrollmentJob (Maybe Text)

The idempotency token for starting a new speaker enrollment Job. If not provided, Amazon Web Services SDK populates this field.

startSpeakerEnrollmentJob_jobName :: Lens' StartSpeakerEnrollmentJob (Maybe Text)

A name for your speaker enrollment job.

startSpeakerEnrollmentJob_enrollmentConfig :: Lens' StartSpeakerEnrollmentJob (Maybe EnrollmentConfig)

The enrollment config that contains details such as the action to take when a speaker is already enrolled in the Voice ID system or when a speaker is identified as a fraudster.

startSpeakerEnrollmentJob_dataAccessRoleArn :: Lens' StartSpeakerEnrollmentJob Text

The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the job output file. Refer to Batch enrollment using audio data from prior calls documentation for the permissions needed in this role.

startSpeakerEnrollmentJob_domainId :: Lens' StartSpeakerEnrollmentJob Text

The identifier of the domain that contains the speaker enrollment job and in which the speakers are enrolled.

startSpeakerEnrollmentJob_inputDataConfig :: Lens' StartSpeakerEnrollmentJob InputDataConfig

The input data config containing the S3 location for the input manifest file that contains the list of speaker enrollment requests.

startSpeakerEnrollmentJob_outputDataConfig :: Lens' StartSpeakerEnrollmentJob OutputDataConfig

The output data config containing the S3 location where Voice ID writes the job output file; you must also include a KMS Key ID to encrypt the file.

startSpeakerEnrollmentJobResponse_job :: Lens' StartSpeakerEnrollmentJobResponse (Maybe SpeakerEnrollmentJob)

Details about the started speaker enrollment job.

startSpeakerEnrollmentJobResponse_httpStatus :: Lens' StartSpeakerEnrollmentJobResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the Voice ID resource you want to remove tags from.

untagResource_tagKeys :: Lens' UntagResource [Text]

The list of tag keys you want to remove from the specified resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

OptOutSpeaker

optOutSpeaker_domainId :: Lens' OptOutSpeaker Text

The identifier of the domain containing the speaker.

optOutSpeaker_speakerId :: Lens' OptOutSpeaker Text

The identifier of the speaker you want opted-out.

optOutSpeakerResponse_speaker :: Lens' OptOutSpeakerResponse (Maybe Speaker)

Details about the opted-out speaker.

optOutSpeakerResponse_httpStatus :: Lens' OptOutSpeakerResponse Int

The response's http status code.

DescribeSpeaker

describeSpeaker_domainId :: Lens' DescribeSpeaker Text

The identifier of the domain that contains the speaker.

describeSpeaker_speakerId :: Lens' DescribeSpeaker Text

The identifier of the speaker you are describing.

describeSpeakerResponse_speaker :: Lens' DescribeSpeakerResponse (Maybe Speaker)

Information about the specified speaker.

describeSpeakerResponse_httpStatus :: Lens' DescribeSpeakerResponse Int

The response's http status code.

DeleteDomain

deleteDomain_domainId :: Lens' DeleteDomain Text

The identifier of the domain you want to delete.

UpdateDomain

updateDomain_description :: Lens' UpdateDomain (Maybe Text)

A brief description about this domain.

updateDomain_domainId :: Lens' UpdateDomain Text

The identifier of the domain to be updated.

updateDomain_name :: Lens' UpdateDomain Text

The name of the domain.

updateDomain_serverSideEncryptionConfiguration :: Lens' UpdateDomain ServerSideEncryptionConfiguration

The configuration, containing the KMS Key Identifier, to be used by Voice ID for the server-side encryption of your data. Note that all the existing data in the domain are still encrypted using the existing key, only the data added to domain after updating the key is encrypted using the new key.

updateDomainResponse_domain :: Lens' UpdateDomainResponse (Maybe Domain)

Details about the updated domain

updateDomainResponse_httpStatus :: Lens' UpdateDomainResponse Int

The response's http status code.

ListDomains

listDomains_nextToken :: Lens' ListDomains (Maybe Text)

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

listDomains_maxResults :: Lens' ListDomains (Maybe Natural)

The maximum number of results that are returned per call. You can use NextToken to obtain further pages of results. The default is 100; the maximum allowed page size is also 100.

listDomainsResponse_nextToken :: Lens' ListDomainsResponse (Maybe Text)

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

listDomainsResponse_domainSummaries :: Lens' ListDomainsResponse (Maybe [DomainSummary])

A list containing details about each domain in the Amazon Web Services account.

listDomainsResponse_httpStatus :: Lens' ListDomainsResponse Int

The response's http status code.

Types

AuthenticationConfiguration

authenticationConfiguration_acceptanceThreshold :: Lens' AuthenticationConfiguration Natural

The minimum threshold needed to successfully authenticate a speaker.

AuthenticationResult

authenticationResult_customerSpeakerId :: Lens' AuthenticationResult (Maybe Text)

The client-provided identifier for the speaker whose authentication result is produced. Only present if a SpeakerId is provided for the session.

authenticationResult_score :: Lens' AuthenticationResult (Maybe Natural)

The authentication score for the speaker whose authentication result is produced. This value is only present if the authentication decision is either ACCEPT or REJECT.

authenticationResult_authenticationResultId :: Lens' AuthenticationResult (Maybe Text)

The unique identifier for this authentication result. Because there can be multiple authentications for a given session, this field helps to identify if the returned result is from a previous streaming activity or a new result. Note that in absence of any new streaming activity, AcceptanceThreshold changes, or SpeakerId changes, Voice ID always returns cached Authentication Result for this API.

authenticationResult_decision :: Lens' AuthenticationResult (Maybe AuthenticationDecision)

The authentication decision produced by Voice ID, processed against the current session state and streamed audio of the speaker.

authenticationResult_configuration :: Lens' AuthenticationResult (Maybe AuthenticationConfiguration)

The AuthenticationConfiguration used to generate this authentication result.

authenticationResult_audioAggregationStartedAt :: Lens' AuthenticationResult (Maybe UTCTime)

A timestamp indicating when audio aggregation started for this authentication result.

authenticationResult_generatedSpeakerId :: Lens' AuthenticationResult (Maybe Text)

The service-generated identifier for the speaker whose authentication result is produced.

authenticationResult_audioAggregationEndedAt :: Lens' AuthenticationResult (Maybe UTCTime)

A timestamp indicating when audio aggregation ended for this authentication result.

Domain

domain_domainStatus :: Lens' Domain (Maybe DomainStatus)

The current status of the domain.

domain_arn :: Lens' Domain (Maybe Text)

The Amazon Resource Name (ARN) for the domain.

domain_createdAt :: Lens' Domain (Maybe UTCTime)

The timestamp at which the domain is created.

domain_name :: Lens' Domain (Maybe Text)

The client-provided name for the domain.

domain_domainId :: Lens' Domain (Maybe Text)

The service-generated identifier for the domain.

domain_updatedAt :: Lens' Domain (Maybe UTCTime)

The timestamp showing the domain's last update.

domain_description :: Lens' Domain (Maybe Text)

The client-provided description of the domain.

domain_serverSideEncryptionConfiguration :: Lens' Domain (Maybe ServerSideEncryptionConfiguration)

The server-side encryption configuration containing the KMS Key Identifier you want Voice ID to use to encrypt your data.

DomainSummary

domainSummary_domainStatus :: Lens' DomainSummary (Maybe DomainStatus)

The current status of the domain.

domainSummary_arn :: Lens' DomainSummary (Maybe Text)

The Amazon Resource Name (ARN) for the domain.

domainSummary_createdAt :: Lens' DomainSummary (Maybe UTCTime)

The timestamp showing when the domain is created.

domainSummary_name :: Lens' DomainSummary (Maybe Text)

The client-provided name for the domain.

domainSummary_domainId :: Lens' DomainSummary (Maybe Text)

The service-generated identifier for the domain.

domainSummary_updatedAt :: Lens' DomainSummary (Maybe UTCTime)

The timestamp showing the domain's last update.

domainSummary_description :: Lens' DomainSummary (Maybe Text)

The client-provided description of the domain.

domainSummary_serverSideEncryptionConfiguration :: Lens' DomainSummary (Maybe ServerSideEncryptionConfiguration)

The server-side encryption configuration containing the KMS Key Identifier you want Voice ID to use to encrypt your data..

EnrollmentConfig

enrollmentConfig_fraudDetectionConfig :: Lens' EnrollmentConfig (Maybe EnrollmentJobFraudDetectionConfig)

The fraud detection configuration to use for the speaker enrollment job.

enrollmentConfig_existingEnrollmentAction :: Lens' EnrollmentConfig (Maybe ExistingEnrollmentAction)

The action to take when the specified speaker is already enrolled in the specified domain. The default value is SKIP, which skips the enrollment for the existing speaker. Setting the value to OVERWRITE replaces the existing voice prints and enrollment audio stored for that speaker with new data generated from the latest audio.

EnrollmentJobFraudDetectionConfig

enrollmentJobFraudDetectionConfig_riskThreshold :: Lens' EnrollmentJobFraudDetectionConfig (Maybe Natural)

Threshold value for determining whether the speaker is a high risk to be fraudulent. If the detected risk score calculated by Voice ID is greater than or equal to the threshold, the speaker is considered a fraudster.

enrollmentJobFraudDetectionConfig_fraudDetectionAction :: Lens' EnrollmentJobFraudDetectionConfig (Maybe FraudDetectionAction)

The action to take when the given speaker is flagged by the fraud detection system. The default value is FAIL, which fails the speaker enrollment. Changing this value to IGNORE results in the speaker being enrolled even if they are flagged by the fraud detection system.

FailureDetails

failureDetails_message :: Lens' FailureDetails (Maybe Text)

A description of the error that caused the batch job failure.

failureDetails_statusCode :: Lens' FailureDetails (Maybe Int)

An HTTP status code representing the nature of the error.

FraudDetectionConfiguration

fraudDetectionConfiguration_riskThreshold :: Lens' FraudDetectionConfiguration Natural

Threshold value for determining whether the speaker is a fraudster. If the detected risk score calculated by Voice ID is higher than the threshold, the speaker is considered a fraudster.

FraudDetectionResult

fraudDetectionResult_reasons :: Lens' FraudDetectionResult (Maybe [FraudDetectionReason])

The reason speaker was flagged by the fraud detection system. This is only be populated if fraud detection Decision is HIGH_RISK, and only has one possible value: KNOWN_FRAUDSTER.

fraudDetectionResult_riskDetails :: Lens' FraudDetectionResult (Maybe FraudRiskDetails)

Details about each risk analyzed for this speaker.

fraudDetectionResult_fraudDetectionResultId :: Lens' FraudDetectionResult (Maybe Text)

The unique identifier for this fraud detection result. Given there can be multiple fraud detections for a given session, this field helps in identifying if the returned result is from previous streaming activity or a new result. Note that in the absence of any new streaming activity or risk threshold changes, Voice ID always returns cached Fraud Detection result for this API.

fraudDetectionResult_decision :: Lens' FraudDetectionResult (Maybe FraudDetectionDecision)

The fraud detection decision produced by Voice ID, processed against the current session state and streamed audio of the speaker.

fraudDetectionResult_configuration :: Lens' FraudDetectionResult (Maybe FraudDetectionConfiguration)

The FraudDetectionConfiguration used to generate this fraud detection result.

fraudDetectionResult_audioAggregationStartedAt :: Lens' FraudDetectionResult (Maybe UTCTime)

A timestamp indicating when audio aggregation started for this fraud detection result.

fraudDetectionResult_audioAggregationEndedAt :: Lens' FraudDetectionResult (Maybe UTCTime)

A timestamp indicating when audio aggregation ended for this fraud detection result.

FraudRiskDetails

fraudRiskDetails_knownFraudsterRisk :: Lens' FraudRiskDetails KnownFraudsterRisk

The details resulting from 'Known Fraudster Risk' analysis of the speaker.

Fraudster

fraudster_createdAt :: Lens' Fraudster (Maybe UTCTime)

The timestamp when Voice ID identified the fraudster.

fraudster_generatedFraudsterId :: Lens' Fraudster (Maybe Text)

The service-generated identifier for the fraudster.

fraudster_domainId :: Lens' Fraudster (Maybe Text)

The identifier for the domain containing the fraudster.

FraudsterRegistrationJob

fraudsterRegistrationJob_failureDetails :: Lens' FraudsterRegistrationJob (Maybe FailureDetails)

Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

fraudsterRegistrationJob_jobId :: Lens' FraudsterRegistrationJob (Maybe Text)

The service-generated identifier for the fraudster registration job.

fraudsterRegistrationJob_createdAt :: Lens' FraudsterRegistrationJob (Maybe UTCTime)

A timestamp showing the creation time of the fraudster registration job.

fraudsterRegistrationJob_jobName :: Lens' FraudsterRegistrationJob (Maybe Text)

The client-provied name for the fraudster registration job.

fraudsterRegistrationJob_endedAt :: Lens' FraudsterRegistrationJob (Maybe UTCTime)

A timestamp showing when the fraudster registration job ended.

fraudsterRegistrationJob_inputDataConfig :: Lens' FraudsterRegistrationJob (Maybe InputDataConfig)

The input data config containing an S3 URI for the input manifest file that contains the list of fraudster registration job requests.

fraudsterRegistrationJob_jobProgress :: Lens' FraudsterRegistrationJob (Maybe JobProgress)

Shows the completed percentage of registration requests listed in the input file.

fraudsterRegistrationJob_outputDataConfig :: Lens' FraudsterRegistrationJob (Maybe OutputDataConfig)

The output data config containing the S3 location where you want Voice ID to write your job output file; you must also include a KMS Key ID in order to encrypt the file.

fraudsterRegistrationJob_dataAccessRoleArn :: Lens' FraudsterRegistrationJob (Maybe Text)

The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the job output file.

fraudsterRegistrationJob_domainId :: Lens' FraudsterRegistrationJob (Maybe Text)

The identifier of the domain containing the fraudster registration job.

fraudsterRegistrationJob_jobStatus :: Lens' FraudsterRegistrationJob (Maybe FraudsterRegistrationJobStatus)

The current status of the fraudster registration job.

fraudsterRegistrationJob_registrationConfig :: Lens' FraudsterRegistrationJob (Maybe RegistrationConfig)

The registration config containing details such as the action to take when a duplicate fraudster is detected, and the similarity threshold to use for detecting a duplicate fraudster.

FraudsterRegistrationJobSummary

fraudsterRegistrationJobSummary_failureDetails :: Lens' FraudsterRegistrationJobSummary (Maybe FailureDetails)

Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

fraudsterRegistrationJobSummary_jobId :: Lens' FraudsterRegistrationJobSummary (Maybe Text)

The service-generated identifier for the fraudster registration job.

fraudsterRegistrationJobSummary_createdAt :: Lens' FraudsterRegistrationJobSummary (Maybe UTCTime)

A timestamp showing when the fraudster registration job is created.

fraudsterRegistrationJobSummary_jobName :: Lens' FraudsterRegistrationJobSummary (Maybe Text)

The client-provied name for the fraudster registration job.

fraudsterRegistrationJobSummary_endedAt :: Lens' FraudsterRegistrationJobSummary (Maybe UTCTime)

A timestamp showing when the fraudster registration job ended.

fraudsterRegistrationJobSummary_jobProgress :: Lens' FraudsterRegistrationJobSummary (Maybe JobProgress)

Shows the completed percentage of registration requests listed in the input file.

fraudsterRegistrationJobSummary_domainId :: Lens' FraudsterRegistrationJobSummary (Maybe Text)

The identifier of the domain containing the fraudster registration job.

fraudsterRegistrationJobSummary_jobStatus :: Lens' FraudsterRegistrationJobSummary (Maybe FraudsterRegistrationJobStatus)

The current status of the fraudster registration job.

InputDataConfig

inputDataConfig_s3Uri :: Lens' InputDataConfig Text

The S3 location for the input manifest file that contains the list of individual enrollment or registration job requests.

JobProgress

jobProgress_percentComplete :: Lens' JobProgress (Maybe Natural)

Shows the completed percentage of enrollment or registration requests listed in the input file.

KnownFraudsterRisk

knownFraudsterRisk_generatedFraudsterId :: Lens' KnownFraudsterRisk (Maybe Text)

The identifier of the fraudster that is the closest match to the speaker. If there are no fraudsters registered in a given domain, or if there are no fraudsters with a non-zero RiskScore, this value is null.

knownFraudsterRisk_riskScore :: Lens' KnownFraudsterRisk Natural

The score indicating the likelihood the speaker is a known fraudster.

OutputDataConfig

outputDataConfig_kmsKeyId :: Lens' OutputDataConfig (Maybe Text)

the identifier of the KMS key you want Voice ID to use to encrypt the output file of the fraudster registration job.

outputDataConfig_s3Uri :: Lens' OutputDataConfig Text

The S3 path of the folder to which Voice ID writes the job output file, which has a *.out extension. For example, if the input file name is input-file.json and the output folder path is s3://output-bucket/output-folder, the full output file path is s3://output-bucket/output-folder/job-Id/input-file.json.out.

RegistrationConfig

registrationConfig_duplicateRegistrationAction :: Lens' RegistrationConfig (Maybe DuplicateRegistrationAction)

The action to take when a fraudster is identified as a duplicate. The default action is SKIP, which skips registering the duplicate fraudster. Setting the value to REGISTER_AS_NEW always registers a new fraudster into the specified domain.

registrationConfig_fraudsterSimilarityThreshold :: Lens' RegistrationConfig (Maybe Natural)

The minimum similarity score between the new and old fraudsters in order to consider the new fraudster a duplicate.

ServerSideEncryptionConfiguration

serverSideEncryptionConfiguration_kmsKeyId :: Lens' ServerSideEncryptionConfiguration Text

The identifier of the KMS Key you want Voice ID to use to encrypt your data.

Speaker

speaker_status :: Lens' Speaker (Maybe SpeakerStatus)

The current status of the speaker.

speaker_customerSpeakerId :: Lens' Speaker (Maybe Text)

The client-provided identifier for the speaker.

speaker_createdAt :: Lens' Speaker (Maybe UTCTime)

A timestamp showing when the speaker is created.

speaker_domainId :: Lens' Speaker (Maybe Text)

The identifier of the domain that contains the speaker.

speaker_updatedAt :: Lens' Speaker (Maybe UTCTime)

A timestamp showing the speaker's last update.

speaker_generatedSpeakerId :: Lens' Speaker (Maybe Text)

The service-generated identifier for the speaker.

SpeakerEnrollmentJob

speakerEnrollmentJob_failureDetails :: Lens' SpeakerEnrollmentJob (Maybe FailureDetails)

Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

speakerEnrollmentJob_jobId :: Lens' SpeakerEnrollmentJob (Maybe Text)

The service-generated identifier for the speaker enrollment job.

speakerEnrollmentJob_createdAt :: Lens' SpeakerEnrollmentJob (Maybe UTCTime)

A timestamp showing the creation of the speaker enrollment job.

speakerEnrollmentJob_jobName :: Lens' SpeakerEnrollmentJob (Maybe Text)

The client-provided name for the speaker enrollment job.

speakerEnrollmentJob_endedAt :: Lens' SpeakerEnrollmentJob (Maybe UTCTime)

A timestamp showing when the speaker enrollment job ended.

speakerEnrollmentJob_enrollmentConfig :: Lens' SpeakerEnrollmentJob (Maybe EnrollmentConfig)

The configuration that defines the action to take when the speaker is already enrolled in Voice ID, and the FraudDetectionConfig to use.

speakerEnrollmentJob_inputDataConfig :: Lens' SpeakerEnrollmentJob (Maybe InputDataConfig)

The input data config containing an S3 URI for the input manifest file that contains the list of speaker enrollment job requests.

speakerEnrollmentJob_jobProgress :: Lens' SpeakerEnrollmentJob (Maybe JobProgress)

Provides details on job progress. This field shows the completed percentage of registration requests listed in the input file.

speakerEnrollmentJob_outputDataConfig :: Lens' SpeakerEnrollmentJob (Maybe OutputDataConfig)

The output data config containing the S3 location where Voice ID writes the job output file; you must also include a KMS Key ID to encrypt the file.

speakerEnrollmentJob_dataAccessRoleArn :: Lens' SpeakerEnrollmentJob (Maybe Text)

The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the job output file.

speakerEnrollmentJob_domainId :: Lens' SpeakerEnrollmentJob (Maybe Text)

The identifier of the domain that contains the speaker enrollment job.

speakerEnrollmentJob_jobStatus :: Lens' SpeakerEnrollmentJob (Maybe SpeakerEnrollmentJobStatus)

The current status of the speaker enrollment job.

SpeakerEnrollmentJobSummary

speakerEnrollmentJobSummary_failureDetails :: Lens' SpeakerEnrollmentJobSummary (Maybe FailureDetails)

Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

speakerEnrollmentJobSummary_jobId :: Lens' SpeakerEnrollmentJobSummary (Maybe Text)

The service-generated identifier for the speaker enrollment job.

speakerEnrollmentJobSummary_createdAt :: Lens' SpeakerEnrollmentJobSummary (Maybe UTCTime)

A timestamp showing the creation time of the speaker enrollment job.

speakerEnrollmentJobSummary_jobName :: Lens' SpeakerEnrollmentJobSummary (Maybe Text)

The client-provided name for the speaker enrollment job.

speakerEnrollmentJobSummary_endedAt :: Lens' SpeakerEnrollmentJobSummary (Maybe UTCTime)

A timestamp showing when the speaker enrollment job ended.

speakerEnrollmentJobSummary_jobProgress :: Lens' SpeakerEnrollmentJobSummary (Maybe JobProgress)

Provides details regarding job progress. This field shows the completed percentage of enrollment requests listed in the input file.

speakerEnrollmentJobSummary_domainId :: Lens' SpeakerEnrollmentJobSummary (Maybe Text)

The identifier of the domain that contains the speaker enrollment job.

speakerEnrollmentJobSummary_jobStatus :: Lens' SpeakerEnrollmentJobSummary (Maybe SpeakerEnrollmentJobStatus)

The current status of the speaker enrollment job.

SpeakerSummary

speakerSummary_status :: Lens' SpeakerSummary (Maybe SpeakerStatus)

The current status of the speaker.

speakerSummary_customerSpeakerId :: Lens' SpeakerSummary (Maybe Text)

The client-provided identifier for the speaker.

speakerSummary_createdAt :: Lens' SpeakerSummary (Maybe UTCTime)

A timestamp showing the speaker's creation time.

speakerSummary_domainId :: Lens' SpeakerSummary (Maybe Text)

The identifier of the domain that contains the speaker.

speakerSummary_updatedAt :: Lens' SpeakerSummary (Maybe UTCTime)

A timestamp showing the speaker's last update.

speakerSummary_generatedSpeakerId :: Lens' SpeakerSummary (Maybe Text)

The service-generated identifier for the speaker.

Tag

tag_key :: Lens' Tag Text

The first part of a key:value pair that forms a tag associated with a given resource. For example, in the tag ‘Department’:’Sales’, the key is 'Department'.

tag_value :: Lens' Tag Text

The second part of a key:value pair that forms a tag associated with a given resource. For example, in the tag ‘Department’:’Sales’, the value is 'Sales'.