Amazonka.VoiceId.ListFraudsterRegistrationJobs

Description

Lists all the fraudster registration jobs in the domain with the given JobStatus. If JobStatus is not provided, this lists all fraudster registration jobs in the given domain.

Creating a Request

data ListFraudsterRegistrationJobs

See: newListFraudsterRegistrationJobs smart constructor.

Constructors

ListFraudsterRegistrationJobs'

Fields

nextToken :: Maybe Text

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

jobStatus :: Maybe FraudsterRegistrationJobStatus

Provides the status of your fraudster registration job.

maxResults :: Maybe Natural

The maximum number of results that are returned per call. You can use NextToken to obtain further pages of results. The default is 100; the maximum allowed page size is also 100.

domainId :: Text

The identifier of the domain containing the fraudster registration Jobs.

newListFraudsterRegistrationJobs

Arguments

:: Text
$sel:domainId:ListFraudsterRegistrationJobs'

-> ListFraudsterRegistrationJobs

Create a value of ListFraudsterRegistrationJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFraudsterRegistrationJobs', listFraudsterRegistrationJobs_nextToken - If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

$sel:jobStatus:ListFraudsterRegistrationJobs', listFraudsterRegistrationJobs_jobStatus - Provides the status of your fraudster registration job.

$sel:maxResults:ListFraudsterRegistrationJobs', listFraudsterRegistrationJobs_maxResults - The maximum number of results that are returned per call. You can use NextToken to obtain further pages of results. The default is 100; the maximum allowed page size is also 100.

$sel:domainId:ListFraudsterRegistrationJobs', listFraudsterRegistrationJobs_domainId - The identifier of the domain containing the fraudster registration Jobs.

Request Lenses

listFraudsterRegistrationJobs_nextToken :: Lens' ListFraudsterRegistrationJobs (Maybe Text)

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

listFraudsterRegistrationJobs_jobStatus :: Lens' ListFraudsterRegistrationJobs (Maybe FraudsterRegistrationJobStatus)

Provides the status of your fraudster registration job.

listFraudsterRegistrationJobs_maxResults :: Lens' ListFraudsterRegistrationJobs (Maybe Natural)

The maximum number of results that are returned per call. You can use NextToken to obtain further pages of results. The default is 100; the maximum allowed page size is also 100.

listFraudsterRegistrationJobs_domainId :: Lens' ListFraudsterRegistrationJobs Text

The identifier of the domain containing the fraudster registration Jobs.

Destructuring the Response

data ListFraudsterRegistrationJobsResponse

See: newListFraudsterRegistrationJobsResponse smart constructor.

Constructors

ListFraudsterRegistrationJobsResponse'

Fields

nextToken :: Maybe Text

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

jobSummaries :: Maybe [FraudsterRegistrationJobSummary]

A list containing details about each specified fraudster registration job.

httpStatus :: Int

The response's http status code.

newListFraudsterRegistrationJobsResponse

Arguments

:: Int
$sel:httpStatus:ListFraudsterRegistrationJobsResponse'

-> ListFraudsterRegistrationJobsResponse

Create a value of ListFraudsterRegistrationJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFraudsterRegistrationJobs', listFraudsterRegistrationJobsResponse_nextToken - If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

$sel:jobSummaries:ListFraudsterRegistrationJobsResponse', listFraudsterRegistrationJobsResponse_jobSummaries - A list containing details about each specified fraudster registration job.

$sel:httpStatus:ListFraudsterRegistrationJobsResponse', listFraudsterRegistrationJobsResponse_httpStatus - The response's http status code.

Response Lenses

listFraudsterRegistrationJobsResponse_nextToken :: Lens' ListFraudsterRegistrationJobsResponse (Maybe Text)

If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

listFraudsterRegistrationJobsResponse_jobSummaries :: Lens' ListFraudsterRegistrationJobsResponse (Maybe [FraudsterRegistrationJobSummary])

A list containing details about each specified fraudster registration job.

listFraudsterRegistrationJobsResponse_httpStatus :: Lens' ListFraudsterRegistrationJobsResponse Int

The response's http status code.