Amazonka.VoiceId.Types.FraudsterRegistrationJob

Description

Documentation

data FraudsterRegistrationJob

Contains all the information about a fraudster registration job.

See: newFraudsterRegistrationJob smart constructor.

Constructors

FraudsterRegistrationJob'

Fields

failureDetails :: Maybe FailureDetails

Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

jobId :: Maybe Text

The service-generated identifier for the fraudster registration job.

createdAt :: Maybe POSIX

A timestamp showing the creation time of the fraudster registration job.

jobName :: Maybe (Sensitive Text)

The client-provied name for the fraudster registration job.

endedAt :: Maybe POSIX

A timestamp showing when the fraudster registration job ended.

inputDataConfig :: Maybe InputDataConfig

The input data config containing an S3 URI for the input manifest file that contains the list of fraudster registration job requests.

jobProgress :: Maybe JobProgress

Shows the completed percentage of registration requests listed in the input file.

outputDataConfig :: Maybe OutputDataConfig

The output data config containing the S3 location where you want Voice ID to write your job output file; you must also include a KMS Key ID in order to encrypt the file.

dataAccessRoleArn :: Maybe Text

The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the job output file.

domainId :: Maybe Text

The identifier of the domain containing the fraudster registration job.

jobStatus :: Maybe FraudsterRegistrationJobStatus

The current status of the fraudster registration job.

registrationConfig :: Maybe RegistrationConfig

The registration config containing details such as the action to take when a duplicate fraudster is detected, and the similarity threshold to use for detecting a duplicate fraudster.

newFraudsterRegistrationJob :: FraudsterRegistrationJob

Create a value of FraudsterRegistrationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureDetails:FraudsterRegistrationJob', fraudsterRegistrationJob_failureDetails - Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

$sel:jobId:FraudsterRegistrationJob', fraudsterRegistrationJob_jobId - The service-generated identifier for the fraudster registration job.

$sel:createdAt:FraudsterRegistrationJob', fraudsterRegistrationJob_createdAt - A timestamp showing the creation time of the fraudster registration job.

$sel:jobName:FraudsterRegistrationJob', fraudsterRegistrationJob_jobName - The client-provied name for the fraudster registration job.

$sel:endedAt:FraudsterRegistrationJob', fraudsterRegistrationJob_endedAt - A timestamp showing when the fraudster registration job ended.

$sel:inputDataConfig:FraudsterRegistrationJob', fraudsterRegistrationJob_inputDataConfig - The input data config containing an S3 URI for the input manifest file that contains the list of fraudster registration job requests.

$sel:jobProgress:FraudsterRegistrationJob', fraudsterRegistrationJob_jobProgress - Shows the completed percentage of registration requests listed in the input file.

$sel:outputDataConfig:FraudsterRegistrationJob', fraudsterRegistrationJob_outputDataConfig - The output data config containing the S3 location where you want Voice ID to write your job output file; you must also include a KMS Key ID in order to encrypt the file.

$sel:dataAccessRoleArn:FraudsterRegistrationJob', fraudsterRegistrationJob_dataAccessRoleArn - The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the job output file.

$sel:domainId:FraudsterRegistrationJob', fraudsterRegistrationJob_domainId - The identifier of the domain containing the fraudster registration job.

$sel:jobStatus:FraudsterRegistrationJob', fraudsterRegistrationJob_jobStatus - The current status of the fraudster registration job.

$sel:registrationConfig:FraudsterRegistrationJob', fraudsterRegistrationJob_registrationConfig - The registration config containing details such as the action to take when a duplicate fraudster is detected, and the similarity threshold to use for detecting a duplicate fraudster.

fraudsterRegistrationJob_failureDetails :: Lens' FraudsterRegistrationJob (Maybe FailureDetails)

Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

fraudsterRegistrationJob_jobId :: Lens' FraudsterRegistrationJob (Maybe Text)

The service-generated identifier for the fraudster registration job.

fraudsterRegistrationJob_createdAt :: Lens' FraudsterRegistrationJob (Maybe UTCTime)

A timestamp showing the creation time of the fraudster registration job.

fraudsterRegistrationJob_jobName :: Lens' FraudsterRegistrationJob (Maybe Text)

The client-provied name for the fraudster registration job.

fraudsterRegistrationJob_endedAt :: Lens' FraudsterRegistrationJob (Maybe UTCTime)

A timestamp showing when the fraudster registration job ended.

fraudsterRegistrationJob_inputDataConfig :: Lens' FraudsterRegistrationJob (Maybe InputDataConfig)

The input data config containing an S3 URI for the input manifest file that contains the list of fraudster registration job requests.

fraudsterRegistrationJob_jobProgress :: Lens' FraudsterRegistrationJob (Maybe JobProgress)

Shows the completed percentage of registration requests listed in the input file.

fraudsterRegistrationJob_outputDataConfig :: Lens' FraudsterRegistrationJob (Maybe OutputDataConfig)

The output data config containing the S3 location where you want Voice ID to write your job output file; you must also include a KMS Key ID in order to encrypt the file.

fraudsterRegistrationJob_dataAccessRoleArn :: Lens' FraudsterRegistrationJob (Maybe Text)

The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the job output file.

fraudsterRegistrationJob_domainId :: Lens' FraudsterRegistrationJob (Maybe Text)

The identifier of the domain containing the fraudster registration job.

fraudsterRegistrationJob_jobStatus :: Lens' FraudsterRegistrationJob (Maybe FraudsterRegistrationJobStatus)

The current status of the fraudster registration job.

fraudsterRegistrationJob_registrationConfig :: Lens' FraudsterRegistrationJob (Maybe RegistrationConfig)

The registration config containing details such as the action to take when a duplicate fraudster is detected, and the similarity threshold to use for detecting a duplicate fraudster.