Amazonka.VoiceId.Types.SpeakerEnrollmentJobSummary

Description

Documentation

data SpeakerEnrollmentJobSummary

Contains a summary of information about a speaker enrollment job.

See: newSpeakerEnrollmentJobSummary smart constructor.

Constructors

SpeakerEnrollmentJobSummary'

Fields

failureDetails :: Maybe FailureDetails

Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

jobId :: Maybe Text

The service-generated identifier for the speaker enrollment job.

createdAt :: Maybe POSIX

A timestamp showing the creation time of the speaker enrollment job.

jobName :: Maybe (Sensitive Text)

The client-provided name for the speaker enrollment job.

endedAt :: Maybe POSIX

A timestamp showing when the speaker enrollment job ended.

jobProgress :: Maybe JobProgress

Provides details regarding job progress. This field shows the completed percentage of enrollment requests listed in the input file.

domainId :: Maybe Text

The identifier of the domain that contains the speaker enrollment job.

jobStatus :: Maybe SpeakerEnrollmentJobStatus

The current status of the speaker enrollment job.

newSpeakerEnrollmentJobSummary :: SpeakerEnrollmentJobSummary

Create a value of SpeakerEnrollmentJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureDetails:SpeakerEnrollmentJobSummary', speakerEnrollmentJobSummary_failureDetails - Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

$sel:jobId:SpeakerEnrollmentJobSummary', speakerEnrollmentJobSummary_jobId - The service-generated identifier for the speaker enrollment job.

$sel:createdAt:SpeakerEnrollmentJobSummary', speakerEnrollmentJobSummary_createdAt - A timestamp showing the creation time of the speaker enrollment job.

$sel:jobName:SpeakerEnrollmentJobSummary', speakerEnrollmentJobSummary_jobName - The client-provided name for the speaker enrollment job.

$sel:endedAt:SpeakerEnrollmentJobSummary', speakerEnrollmentJobSummary_endedAt - A timestamp showing when the speaker enrollment job ended.

$sel:jobProgress:SpeakerEnrollmentJobSummary', speakerEnrollmentJobSummary_jobProgress - Provides details regarding job progress. This field shows the completed percentage of enrollment requests listed in the input file.

$sel:domainId:SpeakerEnrollmentJobSummary', speakerEnrollmentJobSummary_domainId - The identifier of the domain that contains the speaker enrollment job.

$sel:jobStatus:SpeakerEnrollmentJobSummary', speakerEnrollmentJobSummary_jobStatus - The current status of the speaker enrollment job.

speakerEnrollmentJobSummary_failureDetails :: Lens' SpeakerEnrollmentJobSummary (Maybe FailureDetails)

Contains details that are populated when an entire batch job fails. In cases of individual registration job failures, the batch job as a whole doesn't fail; it is completed with a JobStatus of COMPLETED_WITH_ERRORS. You can use the job output file to identify the individual registration requests that failed.

speakerEnrollmentJobSummary_jobId :: Lens' SpeakerEnrollmentJobSummary (Maybe Text)

The service-generated identifier for the speaker enrollment job.

speakerEnrollmentJobSummary_createdAt :: Lens' SpeakerEnrollmentJobSummary (Maybe UTCTime)

A timestamp showing the creation time of the speaker enrollment job.

speakerEnrollmentJobSummary_jobName :: Lens' SpeakerEnrollmentJobSummary (Maybe Text)

The client-provided name for the speaker enrollment job.

speakerEnrollmentJobSummary_endedAt :: Lens' SpeakerEnrollmentJobSummary (Maybe UTCTime)

A timestamp showing when the speaker enrollment job ended.

speakerEnrollmentJobSummary_jobProgress :: Lens' SpeakerEnrollmentJobSummary (Maybe JobProgress)

Provides details regarding job progress. This field shows the completed percentage of enrollment requests listed in the input file.

speakerEnrollmentJobSummary_domainId :: Lens' SpeakerEnrollmentJobSummary (Maybe Text)

The identifier of the domain that contains the speaker enrollment job.

speakerEnrollmentJobSummary_jobStatus :: Lens' SpeakerEnrollmentJobSummary (Maybe SpeakerEnrollmentJobStatus)

The current status of the speaker enrollment job.