Amazonka.VoiceId.Types.SpeakerSummary

Description

Documentation

data SpeakerSummary

Contains a summary of information about a speaker.

See: newSpeakerSummary smart constructor.

Constructors

SpeakerSummary'

Fields

status :: Maybe SpeakerStatus

The current status of the speaker.

customerSpeakerId :: Maybe (Sensitive Text)

The client-provided identifier for the speaker.

createdAt :: Maybe POSIX

A timestamp showing the speaker's creation time.

domainId :: Maybe Text

The identifier of the domain that contains the speaker.

updatedAt :: Maybe POSIX

A timestamp showing the speaker's last update.

generatedSpeakerId :: Maybe Text

The service-generated identifier for the speaker.

newSpeakerSummary :: SpeakerSummary

Create a value of SpeakerSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SpeakerSummary', speakerSummary_status - The current status of the speaker.

$sel:customerSpeakerId:SpeakerSummary', speakerSummary_customerSpeakerId - The client-provided identifier for the speaker.

$sel:createdAt:SpeakerSummary', speakerSummary_createdAt - A timestamp showing the speaker's creation time.

$sel:domainId:SpeakerSummary', speakerSummary_domainId - The identifier of the domain that contains the speaker.

$sel:updatedAt:SpeakerSummary', speakerSummary_updatedAt - A timestamp showing the speaker's last update.

$sel:generatedSpeakerId:SpeakerSummary', speakerSummary_generatedSpeakerId - The service-generated identifier for the speaker.

speakerSummary_status :: Lens' SpeakerSummary (Maybe SpeakerStatus)

The current status of the speaker.

speakerSummary_customerSpeakerId :: Lens' SpeakerSummary (Maybe Text)

The client-provided identifier for the speaker.

speakerSummary_createdAt :: Lens' SpeakerSummary (Maybe UTCTime)

A timestamp showing the speaker's creation time.

speakerSummary_domainId :: Lens' SpeakerSummary (Maybe Text)

The identifier of the domain that contains the speaker.

speakerSummary_updatedAt :: Lens' SpeakerSummary (Maybe UTCTime)

A timestamp showing the speaker's last update.

speakerSummary_generatedSpeakerId :: Lens' SpeakerSummary (Maybe Text)

The service-generated identifier for the speaker.