Amazonka.VoiceId.UpdateDomain

Description

Updates the specified domain. This API has clobber behavior, and clears and replaces all attributes. If an optional field, such as 'Description' is not provided, it is removed from the domain.

Creating a Request

data UpdateDomain

See: newUpdateDomain smart constructor.

Constructors

UpdateDomain'

Fields

description :: Maybe (Sensitive Text)

A brief description about this domain.

domainId :: Text

The identifier of the domain to be updated.

name :: Sensitive Text

The name of the domain.

serverSideEncryptionConfiguration :: ServerSideEncryptionConfiguration

The configuration, containing the KMS Key Identifier, to be used by Voice ID for the server-side encryption of your data. Note that all the existing data in the domain are still encrypted using the existing key, only the data added to domain after updating the key is encrypted using the new key.

newUpdateDomain

Arguments

:: Text
$sel:domainId:UpdateDomain'

-> Text
$sel:name:UpdateDomain'

-> ServerSideEncryptionConfiguration
$sel:serverSideEncryptionConfiguration:UpdateDomain'

-> UpdateDomain

Create a value of UpdateDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:UpdateDomain', updateDomain_description - A brief description about this domain.

$sel:domainId:UpdateDomain', updateDomain_domainId - The identifier of the domain to be updated.

$sel:name:UpdateDomain', updateDomain_name - The name of the domain.

$sel:serverSideEncryptionConfiguration:UpdateDomain', updateDomain_serverSideEncryptionConfiguration - The configuration, containing the KMS Key Identifier, to be used by Voice ID for the server-side encryption of your data. Note that all the existing data in the domain are still encrypted using the existing key, only the data added to domain after updating the key is encrypted using the new key.

Request Lenses

updateDomain_description :: Lens' UpdateDomain (Maybe Text)

A brief description about this domain.

updateDomain_domainId :: Lens' UpdateDomain Text

The identifier of the domain to be updated.

updateDomain_name :: Lens' UpdateDomain Text

The name of the domain.

updateDomain_serverSideEncryptionConfiguration :: Lens' UpdateDomain ServerSideEncryptionConfiguration

The configuration, containing the KMS Key Identifier, to be used by Voice ID for the server-side encryption of your data. Note that all the existing data in the domain are still encrypted using the existing key, only the data added to domain after updating the key is encrypted using the new key.

Destructuring the Response

data UpdateDomainResponse

See: newUpdateDomainResponse smart constructor.

Constructors

UpdateDomainResponse'

Fields

domain :: Maybe Domain

Details about the updated domain

httpStatus :: Int

The response's http status code.

newUpdateDomainResponse

Arguments

:: Int
$sel:httpStatus:UpdateDomainResponse'

-> UpdateDomainResponse

Create a value of UpdateDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domain:UpdateDomainResponse', updateDomainResponse_domain - Details about the updated domain

$sel:httpStatus:UpdateDomainResponse', updateDomainResponse_httpStatus - The response's http status code.

Response Lenses

updateDomainResponse_domain :: Lens' UpdateDomainResponse (Maybe Domain)

Details about the updated domain

updateDomainResponse_httpStatus :: Lens' UpdateDomainResponse Int

The response's http status code.