Amazonka.WAF.DeleteRateBasedRule

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Permanently deletes a RateBasedRule. You can't delete a rule if it's still used in any WebACL objects or if it still includes any predicates, such as ByteMatchSet objects.

If you just want to remove a rule from a WebACL, use UpdateWebACL.

To permanently delete a RateBasedRule from AWS WAF, perform the following steps:

Update the RateBasedRule to remove predicates, if any. For more information, see UpdateRateBasedRule.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a DeleteRateBasedRule request.

Submit a DeleteRateBasedRule request.

Creating a Request

data DeleteRateBasedRule

See: newDeleteRateBasedRule smart constructor.

Constructors

DeleteRateBasedRule'

Fields

ruleId :: Text

The RuleId of the RateBasedRule that you want to delete. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newDeleteRateBasedRule

Arguments

:: Text
$sel:ruleId:DeleteRateBasedRule'

-> Text
$sel:changeToken:DeleteRateBasedRule'

-> DeleteRateBasedRule

Create a value of DeleteRateBasedRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleId:DeleteRateBasedRule', deleteRateBasedRule_ruleId - The RuleId of the RateBasedRule that you want to delete. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules.

$sel:changeToken:DeleteRateBasedRule', deleteRateBasedRule_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

deleteRateBasedRule_ruleId :: Lens' DeleteRateBasedRule Text

The RuleId of the RateBasedRule that you want to delete. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules.

deleteRateBasedRule_changeToken :: Lens' DeleteRateBasedRule Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data DeleteRateBasedRuleResponse

See: newDeleteRateBasedRuleResponse smart constructor.

Constructors

DeleteRateBasedRuleResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the DeleteRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newDeleteRateBasedRuleResponse

Arguments

:: Int
$sel:httpStatus:DeleteRateBasedRuleResponse'

-> DeleteRateBasedRuleResponse

Create a value of DeleteRateBasedRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:DeleteRateBasedRule', deleteRateBasedRuleResponse_changeToken - The ChangeToken that you used to submit the DeleteRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:DeleteRateBasedRuleResponse', deleteRateBasedRuleResponse_httpStatus - The response's http status code.

Response Lenses

deleteRateBasedRuleResponse_changeToken :: Lens' DeleteRateBasedRuleResponse (Maybe Text)

The ChangeToken that you used to submit the DeleteRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

deleteRateBasedRuleResponse_httpStatus :: Lens' DeleteRateBasedRuleResponse Int

The response's http status code.