Amazonka.WAF.Types.HTTPRequest

Description

Documentation

data HTTPRequest

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

The response from a GetSampledRequests request includes an HTTPRequest complex type that appears as Request in the response syntax. HTTPRequest contains information about one of the web requests that were returned by GetSampledRequests.

See: newHTTPRequest smart constructor.

Constructors

HTTPRequest'

Fields

hTTPVersion :: Maybe Text

The HTTP version specified in the sampled web request, for example, HTTP/1.1.

country :: Maybe Text

The two-letter country code for the country that the request originated from. For a current list of country codes, see the Wikipedia entry ISO 3166-1 alpha-2.

uri :: Maybe Text

The part of a web request that identifies the resource, for example, /images/daily-ad.jpg.

headers :: Maybe [HTTPHeader]

A complex type that contains two values for each header in the sampled web request: the name of the header and the value of the header.

method :: Maybe Text

The HTTP method specified in the sampled web request. CloudFront supports the following methods: DELETE, GET, HEAD, OPTIONS, PATCH, POST, and PUT.

clientIP :: Maybe Text

The IP address that the request originated from. If the WebACL is associated with a CloudFront distribution, this is the value of one of the following fields in CloudFront access logs:

c-ip, if the viewer did not use an HTTP proxy or a load balancer to send the request

x-forwarded-for, if the viewer did use an HTTP proxy or a load balancer to send the request

newHTTPRequest :: HTTPRequest

Create a value of HTTPRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hTTPVersion:HTTPRequest', hTTPRequest_hTTPVersion - The HTTP version specified in the sampled web request, for example, HTTP/1.1.

$sel:country:HTTPRequest', hTTPRequest_country - The two-letter country code for the country that the request originated from. For a current list of country codes, see the Wikipedia entry ISO 3166-1 alpha-2.

$sel:uri:HTTPRequest', hTTPRequest_uri - The part of a web request that identifies the resource, for example, /images/daily-ad.jpg.

$sel:headers:HTTPRequest', hTTPRequest_headers - A complex type that contains two values for each header in the sampled web request: the name of the header and the value of the header.

$sel:method:HTTPRequest', hTTPRequest_method - The HTTP method specified in the sampled web request. CloudFront supports the following methods: DELETE, GET, HEAD, OPTIONS, PATCH, POST, and PUT.

$sel:clientIP:HTTPRequest', hTTPRequest_clientIP - The IP address that the request originated from. If the WebACL is associated with a CloudFront distribution, this is the value of one of the following fields in CloudFront access logs:

c-ip, if the viewer did not use an HTTP proxy or a load balancer to send the request

x-forwarded-for, if the viewer did use an HTTP proxy or a load balancer to send the request

hTTPRequest_hTTPVersion :: Lens' HTTPRequest (Maybe Text)

The HTTP version specified in the sampled web request, for example, HTTP/1.1.

hTTPRequest_country :: Lens' HTTPRequest (Maybe Text)

The two-letter country code for the country that the request originated from. For a current list of country codes, see the Wikipedia entry ISO 3166-1 alpha-2.

hTTPRequest_uri :: Lens' HTTPRequest (Maybe Text)

The part of a web request that identifies the resource, for example, /images/daily-ad.jpg.

hTTPRequest_headers :: Lens' HTTPRequest (Maybe [HTTPHeader])

A complex type that contains two values for each header in the sampled web request: the name of the header and the value of the header.

hTTPRequest_method :: Lens' HTTPRequest (Maybe Text)

The HTTP method specified in the sampled web request. CloudFront supports the following methods: DELETE, GET, HEAD, OPTIONS, PATCH, POST, and PUT.

hTTPRequest_clientIP :: Lens' HTTPRequest (Maybe Text)

The IP address that the request originated from. If the WebACL is associated with a CloudFront distribution, this is the value of one of the following fields in CloudFront access logs:

c-ip, if the viewer did not use an HTTP proxy or a load balancer to send the request

x-forwarded-for, if the viewer did use an HTTP proxy or a load balancer to send the request