Amazonka.WAF.Types.LoggingConfiguration

Description

Documentation

data LoggingConfiguration

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

The Amazon Kinesis Data Firehose, RedactedFields information, and the web ACL Amazon Resource Name (ARN).

See: newLoggingConfiguration smart constructor.

Constructors

LoggingConfiguration'

Fields

redactedFields :: Maybe [FieldToMatch]

The parts of the request that you want redacted from the logs. For example, if you redact the cookie field, the cookie field in the firehose will be xxx.

resourceArn :: Text

The Amazon Resource Name (ARN) of the web ACL that you want to associate with LogDestinationConfigs.

logDestinationConfigs :: NonEmpty Text

An array of Amazon Kinesis Data Firehose ARNs.

newLoggingConfiguration

Arguments

:: Text
$sel:resourceArn:LoggingConfiguration'

-> NonEmpty Text
$sel:logDestinationConfigs:LoggingConfiguration'

-> LoggingConfiguration

Create a value of LoggingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:redactedFields:LoggingConfiguration', loggingConfiguration_redactedFields - The parts of the request that you want redacted from the logs. For example, if you redact the cookie field, the cookie field in the firehose will be xxx.

$sel:resourceArn:LoggingConfiguration', loggingConfiguration_resourceArn - The Amazon Resource Name (ARN) of the web ACL that you want to associate with LogDestinationConfigs.

$sel:logDestinationConfigs:LoggingConfiguration', loggingConfiguration_logDestinationConfigs - An array of Amazon Kinesis Data Firehose ARNs.

loggingConfiguration_redactedFields :: Lens' LoggingConfiguration (Maybe [FieldToMatch])

The parts of the request that you want redacted from the logs. For example, if you redact the cookie field, the cookie field in the firehose will be xxx.

loggingConfiguration_resourceArn :: Lens' LoggingConfiguration Text

The Amazon Resource Name (ARN) of the web ACL that you want to associate with LogDestinationConfigs.

loggingConfiguration_logDestinationConfigs :: Lens' LoggingConfiguration (NonEmpty Text)

An array of Amazon Kinesis Data Firehose ARNs.