Amazonka.WAF.Types.RuleGroup

Description

Documentation

data RuleGroup

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

A collection of predefined rules that you can add to a web ACL.

Rule groups are subject to the following limits:

Three rule groups per account. You can request an increase to this limit by contacting customer support.

One rule group per web ACL.

Ten rules per rule group.

See: newRuleGroup smart constructor.

Constructors

RuleGroup'

Fields

metricName :: Maybe Text

A friendly name or description for the metrics for this RuleGroup. The name can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum length 128 and minimum length one. It can't contain whitespace or metric names reserved for AWS WAF, including "All" and "Default_Action." You can't change the name of the metric after you create the RuleGroup.

name :: Maybe Text

The friendly name or description for the RuleGroup. You can't change the name of a RuleGroup after you create it.

ruleGroupId :: Text

A unique identifier for a RuleGroup. You use RuleGroupId to get more information about a RuleGroup (see GetRuleGroup), update a RuleGroup (see UpdateRuleGroup), insert a RuleGroup into a WebACL or delete a one from a WebACL (see UpdateWebACL), or delete a RuleGroup from AWS WAF (see DeleteRuleGroup).

RuleGroupId is returned by CreateRuleGroup and by ListRuleGroups.

newRuleGroup

Arguments

:: Text
$sel:ruleGroupId:RuleGroup'

-> RuleGroup

Create a value of RuleGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:RuleGroup', ruleGroup_metricName - A friendly name or description for the metrics for this RuleGroup. The name can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum length 128 and minimum length one. It can't contain whitespace or metric names reserved for AWS WAF, including "All" and "Default_Action." You can't change the name of the metric after you create the RuleGroup.

$sel:name:RuleGroup', ruleGroup_name - The friendly name or description for the RuleGroup. You can't change the name of a RuleGroup after you create it.

$sel:ruleGroupId:RuleGroup', ruleGroup_ruleGroupId - A unique identifier for a RuleGroup. You use RuleGroupId to get more information about a RuleGroup (see GetRuleGroup), update a RuleGroup (see UpdateRuleGroup), insert a RuleGroup into a WebACL or delete a one from a WebACL (see UpdateWebACL), or delete a RuleGroup from AWS WAF (see DeleteRuleGroup).

RuleGroupId is returned by CreateRuleGroup and by ListRuleGroups.

ruleGroup_metricName :: Lens' RuleGroup (Maybe Text)

A friendly name or description for the metrics for this RuleGroup. The name can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum length 128 and minimum length one. It can't contain whitespace or metric names reserved for AWS WAF, including "All" and "Default_Action." You can't change the name of the metric after you create the RuleGroup.

ruleGroup_name :: Lens' RuleGroup (Maybe Text)

The friendly name or description for the RuleGroup. You can't change the name of a RuleGroup after you create it.

ruleGroup_ruleGroupId :: Lens' RuleGroup Text

A unique identifier for a RuleGroup. You use RuleGroupId to get more information about a RuleGroup (see GetRuleGroup), update a RuleGroup (see UpdateRuleGroup), insert a RuleGroup into a WebACL or delete a one from a WebACL (see UpdateWebACL), or delete a RuleGroup from AWS WAF (see DeleteRuleGroup).

RuleGroupId is returned by CreateRuleGroup and by ListRuleGroups.