Amazonka.WAF.Types.WebACLUpdate

Description

Documentation

data WebACLUpdate

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Specifies whether to insert a Rule into or delete a Rule from a WebACL.

See: newWebACLUpdate smart constructor.

Constructors

WebACLUpdate'

Fields

action :: ChangeAction

Specifies whether to insert a Rule into or delete a Rule from a WebACL.

activatedRule :: ActivatedRule

The ActivatedRule object in an UpdateWebACL request specifies a Rule that you want to insert or delete, the priority of the Rule in the WebACL, and the action that you want AWS WAF to take when a web request matches the Rule (ALLOW, BLOCK, or COUNT).

newWebACLUpdate

Arguments

:: ChangeAction
$sel:action:WebACLUpdate'

-> ActivatedRule
$sel:activatedRule:WebACLUpdate'

-> WebACLUpdate

Create a value of WebACLUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:WebACLUpdate', webACLUpdate_action - Specifies whether to insert a Rule into or delete a Rule from a WebACL.

$sel:activatedRule:WebACLUpdate', webACLUpdate_activatedRule - The ActivatedRule object in an UpdateWebACL request specifies a Rule that you want to insert or delete, the priority of the Rule in the WebACL, and the action that you want AWS WAF to take when a web request matches the Rule (ALLOW, BLOCK, or COUNT).

webACLUpdate_action :: Lens' WebACLUpdate ChangeAction

Specifies whether to insert a Rule into or delete a Rule from a WebACL.

webACLUpdate_activatedRule :: Lens' WebACLUpdate ActivatedRule

The ActivatedRule object in an UpdateWebACL request specifies a Rule that you want to insert or delete, the priority of the Rule in the WebACL, and the action that you want AWS WAF to take when a web request matches the Rule (ALLOW, BLOCK, or COUNT).