Amazonka.WAF.UpdateRegexPatternSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Inserts or deletes RegexPatternString objects in a RegexPatternSet. For each RegexPatternString object, you specify the following values:

Whether to insert or delete the RegexPatternString.

The regular expression pattern that you want to insert or delete. For more information, see RegexPatternSet.

For example, you can create a RegexPatternString such as B[a@]dB[o0]t. AWS WAF will match this RegexPatternString to:

BadBot

BadB0t

B@dBot

B@dB0t

To create and configure a RegexPatternSet, perform the following steps:

Create a RegexPatternSet. For more information, see CreateRegexPatternSet.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateRegexPatternSet request.

Submit an UpdateRegexPatternSet request to specify the regular expression pattern that you want AWS WAF to watch for.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data UpdateRegexPatternSet

See: newUpdateRegexPatternSet smart constructor.

Constructors

UpdateRegexPatternSet'

Fields

regexPatternSetId :: Text

The RegexPatternSetId of the RegexPatternSet that you want to update. RegexPatternSetId is returned by CreateRegexPatternSet and by ListRegexPatternSets.

updates :: NonEmpty RegexPatternSetUpdate

An array of RegexPatternSetUpdate objects that you want to insert into or delete from a RegexPatternSet.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

newUpdateRegexPatternSet

Arguments

:: Text
$sel:regexPatternSetId:UpdateRegexPatternSet'

-> NonEmpty RegexPatternSetUpdate
$sel:updates:UpdateRegexPatternSet'

-> Text
$sel:changeToken:UpdateRegexPatternSet'

-> UpdateRegexPatternSet

Create a value of UpdateRegexPatternSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regexPatternSetId:UpdateRegexPatternSet', updateRegexPatternSet_regexPatternSetId - The RegexPatternSetId of the RegexPatternSet that you want to update. RegexPatternSetId is returned by CreateRegexPatternSet and by ListRegexPatternSets.

$sel:updates:UpdateRegexPatternSet', updateRegexPatternSet_updates - An array of RegexPatternSetUpdate objects that you want to insert into or delete from a RegexPatternSet.

$sel:changeToken:UpdateRegexPatternSet', updateRegexPatternSet_changeToken - The value returned by the most recent call to GetChangeToken.

Request Lenses

updateRegexPatternSet_regexPatternSetId :: Lens' UpdateRegexPatternSet Text

The RegexPatternSetId of the RegexPatternSet that you want to update. RegexPatternSetId is returned by CreateRegexPatternSet and by ListRegexPatternSets.

updateRegexPatternSet_updates :: Lens' UpdateRegexPatternSet (NonEmpty RegexPatternSetUpdate)

An array of RegexPatternSetUpdate objects that you want to insert into or delete from a RegexPatternSet.

updateRegexPatternSet_changeToken :: Lens' UpdateRegexPatternSet Text

The value returned by the most recent call to GetChangeToken.

Destructuring the Response

data UpdateRegexPatternSetResponse

See: newUpdateRegexPatternSetResponse smart constructor.

Constructors

UpdateRegexPatternSetResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the UpdateRegexPatternSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newUpdateRegexPatternSetResponse

Arguments

:: Int
$sel:httpStatus:UpdateRegexPatternSetResponse'

-> UpdateRegexPatternSetResponse

Create a value of UpdateRegexPatternSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:UpdateRegexPatternSet', updateRegexPatternSetResponse_changeToken - The ChangeToken that you used to submit the UpdateRegexPatternSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:UpdateRegexPatternSetResponse', updateRegexPatternSetResponse_httpStatus - The response's http status code.

Response Lenses

updateRegexPatternSetResponse_changeToken :: Lens' UpdateRegexPatternSetResponse (Maybe Text)

The ChangeToken that you used to submit the UpdateRegexPatternSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

updateRegexPatternSetResponse_httpStatus :: Lens' UpdateRegexPatternSetResponse Int

The response's http status code.