Amazonka.WAFRegional.GetByteMatchSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Returns the ByteMatchSet specified by ByteMatchSetId.

Creating a Request

data GetByteMatchSet

See: newGetByteMatchSet smart constructor.

Constructors

GetByteMatchSet'

Fields

byteMatchSetId :: Text

The ByteMatchSetId of the ByteMatchSet that you want to get. ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

newGetByteMatchSet

Arguments

:: Text
$sel:byteMatchSetId:GetByteMatchSet'

-> GetByteMatchSet

Create a value of GetByteMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:byteMatchSetId:GetByteMatchSet', getByteMatchSet_byteMatchSetId - The ByteMatchSetId of the ByteMatchSet that you want to get. ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

Request Lenses

getByteMatchSet_byteMatchSetId :: Lens' GetByteMatchSet Text

The ByteMatchSetId of the ByteMatchSet that you want to get. ByteMatchSetId is returned by CreateByteMatchSet and by ListByteMatchSets.

Destructuring the Response

data GetByteMatchSetResponse

See: newGetByteMatchSetResponse smart constructor.

Constructors

GetByteMatchSetResponse'

Fields

byteMatchSet :: Maybe ByteMatchSet

Information about the ByteMatchSet that you specified in the GetByteMatchSet request. For more information, see the following topics:

ByteMatchSet: Contains ByteMatchSetId, ByteMatchTuples, and Name

ByteMatchTuples: Contains an array of ByteMatchTuple objects. Each ByteMatchTuple object contains FieldToMatch, PositionalConstraint, TargetString, and TextTransformation

FieldToMatch: Contains Data and Type

httpStatus :: Int

The response's http status code.

newGetByteMatchSetResponse

Arguments

:: Int
$sel:httpStatus:GetByteMatchSetResponse'

-> GetByteMatchSetResponse

Create a value of GetByteMatchSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:byteMatchSet:GetByteMatchSetResponse', getByteMatchSetResponse_byteMatchSet - Information about the ByteMatchSet that you specified in the GetByteMatchSet request. For more information, see the following topics:

ByteMatchSet: Contains ByteMatchSetId, ByteMatchTuples, and Name

ByteMatchTuples: Contains an array of ByteMatchTuple objects. Each ByteMatchTuple object contains FieldToMatch, PositionalConstraint, TargetString, and TextTransformation

FieldToMatch: Contains Data and Type

$sel:httpStatus:GetByteMatchSetResponse', getByteMatchSetResponse_httpStatus - The response's http status code.

Response Lenses

getByteMatchSetResponse_byteMatchSet :: Lens' GetByteMatchSetResponse (Maybe ByteMatchSet)

Information about the ByteMatchSet that you specified in the GetByteMatchSet request. For more information, see the following topics:

ByteMatchSet: Contains ByteMatchSetId, ByteMatchTuples, and Name

ByteMatchTuples: Contains an array of ByteMatchTuple objects. Each ByteMatchTuple object contains FieldToMatch, PositionalConstraint, TargetString, and TextTransformation

FieldToMatch: Contains Data and Type

getByteMatchSetResponse_httpStatus :: Lens' GetByteMatchSetResponse Int

The response's http status code.