Amazonka.WAFRegional.GetGeoMatchSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Returns the GeoMatchSet that is specified by GeoMatchSetId.

Creating a Request

data GetGeoMatchSet

See: newGetGeoMatchSet smart constructor.

Constructors

GetGeoMatchSet'

Fields

geoMatchSetId :: Text

The GeoMatchSetId of the GeoMatchSet that you want to get. GeoMatchSetId is returned by CreateGeoMatchSet and by ListGeoMatchSets.

newGetGeoMatchSet

Arguments

:: Text
$sel:geoMatchSetId:GetGeoMatchSet'

-> GetGeoMatchSet

Create a value of GetGeoMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:geoMatchSetId:GetGeoMatchSet', getGeoMatchSet_geoMatchSetId - The GeoMatchSetId of the GeoMatchSet that you want to get. GeoMatchSetId is returned by CreateGeoMatchSet and by ListGeoMatchSets.

Request Lenses

getGeoMatchSet_geoMatchSetId :: Lens' GetGeoMatchSet Text

The GeoMatchSetId of the GeoMatchSet that you want to get. GeoMatchSetId is returned by CreateGeoMatchSet and by ListGeoMatchSets.

Destructuring the Response

data GetGeoMatchSetResponse

See: newGetGeoMatchSetResponse smart constructor.

Constructors

GetGeoMatchSetResponse'

Fields

geoMatchSet :: Maybe GeoMatchSet

Information about the GeoMatchSet that you specified in the GetGeoMatchSet request. This includes the Type, which for a GeoMatchContraint is always Country, as well as the Value, which is the identifier for a specific country.

httpStatus :: Int

The response's http status code.

newGetGeoMatchSetResponse

Arguments

:: Int
$sel:httpStatus:GetGeoMatchSetResponse'

-> GetGeoMatchSetResponse

Create a value of GetGeoMatchSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:geoMatchSet:GetGeoMatchSetResponse', getGeoMatchSetResponse_geoMatchSet - Information about the GeoMatchSet that you specified in the GetGeoMatchSet request. This includes the Type, which for a GeoMatchContraint is always Country, as well as the Value, which is the identifier for a specific country.

$sel:httpStatus:GetGeoMatchSetResponse', getGeoMatchSetResponse_httpStatus - The response's http status code.

Response Lenses

getGeoMatchSetResponse_geoMatchSet :: Lens' GetGeoMatchSetResponse (Maybe GeoMatchSet)

Information about the GeoMatchSet that you specified in the GetGeoMatchSet request. This includes the Type, which for a GeoMatchContraint is always Country, as well as the Value, which is the identifier for a specific country.

getGeoMatchSetResponse_httpStatus :: Lens' GetGeoMatchSetResponse Int

The response's http status code.