Amazonka.WAFRegional.ListActivatedRulesInRuleGroup

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Returns an array of ActivatedRule objects.

Creating a Request

data ListActivatedRulesInRuleGroup

See: newListActivatedRulesInRuleGroup smart constructor.

Constructors

ListActivatedRulesInRuleGroup'

Fields

ruleGroupId :: Maybe Text

The RuleGroupId of the RuleGroup for which you want to get a list of ActivatedRule objects.

nextMarker :: Maybe Text

If you specify a value for Limit and you have more ActivatedRules than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of ActivatedRules. For the second and subsequent ListActivatedRulesInRuleGroup requests, specify the value of NextMarker from the previous response to get information about another batch of ActivatedRules.

limit :: Maybe Natural

Specifies the number of ActivatedRules that you want AWS WAF to return for this request. If you have more ActivatedRules than the number that you specify for Limit, the response includes a NextMarker value that you can use to get another batch of ActivatedRules.

newListActivatedRulesInRuleGroup :: ListActivatedRulesInRuleGroup

Create a value of ListActivatedRulesInRuleGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleGroupId:ListActivatedRulesInRuleGroup', listActivatedRulesInRuleGroup_ruleGroupId - The RuleGroupId of the RuleGroup for which you want to get a list of ActivatedRule objects.

$sel:nextMarker:ListActivatedRulesInRuleGroup', listActivatedRulesInRuleGroup_nextMarker - If you specify a value for Limit and you have more ActivatedRules than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of ActivatedRules. For the second and subsequent ListActivatedRulesInRuleGroup requests, specify the value of NextMarker from the previous response to get information about another batch of ActivatedRules.

$sel:limit:ListActivatedRulesInRuleGroup', listActivatedRulesInRuleGroup_limit - Specifies the number of ActivatedRules that you want AWS WAF to return for this request. If you have more ActivatedRules than the number that you specify for Limit, the response includes a NextMarker value that you can use to get another batch of ActivatedRules.

Request Lenses

listActivatedRulesInRuleGroup_ruleGroupId :: Lens' ListActivatedRulesInRuleGroup (Maybe Text)

The RuleGroupId of the RuleGroup for which you want to get a list of ActivatedRule objects.

listActivatedRulesInRuleGroup_nextMarker :: Lens' ListActivatedRulesInRuleGroup (Maybe Text)

If you specify a value for Limit and you have more ActivatedRules than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of ActivatedRules. For the second and subsequent ListActivatedRulesInRuleGroup requests, specify the value of NextMarker from the previous response to get information about another batch of ActivatedRules.

listActivatedRulesInRuleGroup_limit :: Lens' ListActivatedRulesInRuleGroup (Maybe Natural)

Specifies the number of ActivatedRules that you want AWS WAF to return for this request. If you have more ActivatedRules than the number that you specify for Limit, the response includes a NextMarker value that you can use to get another batch of ActivatedRules.

Destructuring the Response

data ListActivatedRulesInRuleGroupResponse

See: newListActivatedRulesInRuleGroupResponse smart constructor.

Constructors

ListActivatedRulesInRuleGroupResponse'

Fields

nextMarker :: Maybe Text

If you have more ActivatedRules than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more ActivatedRules, submit another ListActivatedRulesInRuleGroup request, and specify the NextMarker value from the response in the NextMarker value in the next request.

activatedRules :: Maybe [ActivatedRule]

An array of ActivatedRules objects.

httpStatus :: Int

The response's http status code.

newListActivatedRulesInRuleGroupResponse

Arguments

:: Int
$sel:httpStatus:ListActivatedRulesInRuleGroupResponse'

-> ListActivatedRulesInRuleGroupResponse

Create a value of ListActivatedRulesInRuleGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:ListActivatedRulesInRuleGroup', listActivatedRulesInRuleGroupResponse_nextMarker - If you have more ActivatedRules than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more ActivatedRules, submit another ListActivatedRulesInRuleGroup request, and specify the NextMarker value from the response in the NextMarker value in the next request.

$sel:activatedRules:ListActivatedRulesInRuleGroupResponse', listActivatedRulesInRuleGroupResponse_activatedRules - An array of ActivatedRules objects.

$sel:httpStatus:ListActivatedRulesInRuleGroupResponse', listActivatedRulesInRuleGroupResponse_httpStatus - The response's http status code.

Response Lenses

listActivatedRulesInRuleGroupResponse_nextMarker :: Lens' ListActivatedRulesInRuleGroupResponse (Maybe Text)

If you have more ActivatedRules than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more ActivatedRules, submit another ListActivatedRulesInRuleGroup request, and specify the NextMarker value from the response in the NextMarker value in the next request.

listActivatedRulesInRuleGroupResponse_activatedRules :: Lens' ListActivatedRulesInRuleGroupResponse (Maybe [ActivatedRule])

An array of ActivatedRules objects.

listActivatedRulesInRuleGroupResponse_httpStatus :: Lens' ListActivatedRulesInRuleGroupResponse Int

The response's http status code.