Amazonka.WAFRegional.ListRuleGroups

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Returns an array of RuleGroup objects.

Creating a Request

data ListRuleGroups

See: newListRuleGroups smart constructor.

Constructors

ListRuleGroups'

Fields

nextMarker :: Maybe Text

If you specify a value for Limit and you have more RuleGroups than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of RuleGroups. For the second and subsequent ListRuleGroups requests, specify the value of NextMarker from the previous response to get information about another batch of RuleGroups.

limit :: Maybe Natural

Specifies the number of RuleGroups that you want AWS WAF to return for this request. If you have more RuleGroups than the number that you specify for Limit, the response includes a NextMarker value that you can use to get another batch of RuleGroups.

newListRuleGroups :: ListRuleGroups

Create a value of ListRuleGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:ListRuleGroups', listRuleGroups_nextMarker - If you specify a value for Limit and you have more RuleGroups than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of RuleGroups. For the second and subsequent ListRuleGroups requests, specify the value of NextMarker from the previous response to get information about another batch of RuleGroups.

$sel:limit:ListRuleGroups', listRuleGroups_limit - Specifies the number of RuleGroups that you want AWS WAF to return for this request. If you have more RuleGroups than the number that you specify for Limit, the response includes a NextMarker value that you can use to get another batch of RuleGroups.

Request Lenses

listRuleGroups_nextMarker :: Lens' ListRuleGroups (Maybe Text)

If you specify a value for Limit and you have more RuleGroups than the value of Limit, AWS WAF returns a NextMarker value in the response that allows you to list another group of RuleGroups. For the second and subsequent ListRuleGroups requests, specify the value of NextMarker from the previous response to get information about another batch of RuleGroups.

listRuleGroups_limit :: Lens' ListRuleGroups (Maybe Natural)

Specifies the number of RuleGroups that you want AWS WAF to return for this request. If you have more RuleGroups than the number that you specify for Limit, the response includes a NextMarker value that you can use to get another batch of RuleGroups.

Destructuring the Response

data ListRuleGroupsResponse

See: newListRuleGroupsResponse smart constructor.

Constructors

ListRuleGroupsResponse'

Fields

ruleGroups :: Maybe [RuleGroupSummary]

An array of RuleGroup objects.

nextMarker :: Maybe Text

If you have more RuleGroups than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more RuleGroups, submit another ListRuleGroups request, and specify the NextMarker value from the response in the NextMarker value in the next request.

httpStatus :: Int

The response's http status code.

newListRuleGroupsResponse

Arguments

:: Int
$sel:httpStatus:ListRuleGroupsResponse'

-> ListRuleGroupsResponse

Create a value of ListRuleGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleGroups:ListRuleGroupsResponse', listRuleGroupsResponse_ruleGroups - An array of RuleGroup objects.

$sel:nextMarker:ListRuleGroups', listRuleGroupsResponse_nextMarker - If you have more RuleGroups than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more RuleGroups, submit another ListRuleGroups request, and specify the NextMarker value from the response in the NextMarker value in the next request.

$sel:httpStatus:ListRuleGroupsResponse', listRuleGroupsResponse_httpStatus - The response's http status code.

Response Lenses

listRuleGroupsResponse_ruleGroups :: Lens' ListRuleGroupsResponse (Maybe [RuleGroupSummary])

An array of RuleGroup objects.

listRuleGroupsResponse_nextMarker :: Lens' ListRuleGroupsResponse (Maybe Text)

If you have more RuleGroups than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more RuleGroups, submit another ListRuleGroups request, and specify the NextMarker value from the response in the NextMarker value in the next request.

listRuleGroupsResponse_httpStatus :: Lens' ListRuleGroupsResponse Int

The response's http status code.