Amazonka.WAFRegional.PutLoggingConfiguration

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Associates a LoggingConfiguration with a specified web ACL.

You can access information about all traffic that AWS WAF inspects using the following steps:

Create an Amazon Kinesis Data Firehose.

Create the data firehose with a PUT source and in the region that you are operating. However, if you are capturing logs for Amazon CloudFront, always create the firehose in US East (N. Virginia).

Do not create the data firehose using a Kinesis stream as your source.

Associate that firehose to your web ACL using a PutLoggingConfiguration request.

When you successfully enable logging using a PutLoggingConfiguration request, AWS WAF will create a service linked role with the necessary permissions to write logs to the Amazon Kinesis Data Firehose. For more information, see Logging Web ACL Traffic Information in the AWS WAF Developer Guide.

Creating a Request

data PutLoggingConfiguration

See: newPutLoggingConfiguration smart constructor.

Constructors

PutLoggingConfiguration'

Fields

loggingConfiguration :: LoggingConfiguration

The Amazon Kinesis Data Firehose that contains the inspected traffic information, the redacted fields details, and the Amazon Resource Name (ARN) of the web ACL to monitor.

When specifying Type in RedactedFields, you must use one of the following values: URI, QUERY_STRING, HEADER, or METHOD.

newPutLoggingConfiguration

Arguments

:: LoggingConfiguration
$sel:loggingConfiguration:PutLoggingConfiguration'

-> PutLoggingConfiguration

Create a value of PutLoggingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loggingConfiguration:PutLoggingConfiguration', putLoggingConfiguration_loggingConfiguration - The Amazon Kinesis Data Firehose that contains the inspected traffic information, the redacted fields details, and the Amazon Resource Name (ARN) of the web ACL to monitor.

When specifying Type in RedactedFields, you must use one of the following values: URI, QUERY_STRING, HEADER, or METHOD.

Request Lenses

putLoggingConfiguration_loggingConfiguration :: Lens' PutLoggingConfiguration LoggingConfiguration

The Amazon Kinesis Data Firehose that contains the inspected traffic information, the redacted fields details, and the Amazon Resource Name (ARN) of the web ACL to monitor.

When specifying Type in RedactedFields, you must use one of the following values: URI, QUERY_STRING, HEADER, or METHOD.

Destructuring the Response

data PutLoggingConfigurationResponse

See: newPutLoggingConfigurationResponse smart constructor.

Constructors

PutLoggingConfigurationResponse'

Fields

loggingConfiguration :: Maybe LoggingConfiguration

The LoggingConfiguration that you submitted in the request.

httpStatus :: Int

The response's http status code.

newPutLoggingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:PutLoggingConfigurationResponse'

-> PutLoggingConfigurationResponse

Create a value of PutLoggingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loggingConfiguration:PutLoggingConfiguration', putLoggingConfigurationResponse_loggingConfiguration - The LoggingConfiguration that you submitted in the request.

$sel:httpStatus:PutLoggingConfigurationResponse', putLoggingConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

putLoggingConfigurationResponse_loggingConfiguration :: Lens' PutLoggingConfigurationResponse (Maybe LoggingConfiguration)

The LoggingConfiguration that you submitted in the request.

putLoggingConfigurationResponse_httpStatus :: Lens' PutLoggingConfigurationResponse Int

The response's http status code.