Amazonka.WAFRegional.Types.GeoMatchSet

Description

Documentation

data GeoMatchSet

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Contains one or more countries that AWS WAF will search for.

See: newGeoMatchSet smart constructor.

Constructors

GeoMatchSet'

Fields

name :: Maybe Text

A friendly name or description of the GeoMatchSet. You can't change the name of an GeoMatchSet after you create it.

geoMatchSetId :: Text

The GeoMatchSetId for an GeoMatchSet. You use GeoMatchSetId to get information about a GeoMatchSet (see GeoMatchSet), update a GeoMatchSet (see UpdateGeoMatchSet), insert a GeoMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete a GeoMatchSet from AWS WAF (see DeleteGeoMatchSet).

GeoMatchSetId is returned by CreateGeoMatchSet and by ListGeoMatchSets.

geoMatchConstraints :: [GeoMatchConstraint]

An array of GeoMatchConstraint objects, which contain the country that you want AWS WAF to search for.

newGeoMatchSet

Arguments

:: Text
$sel:geoMatchSetId:GeoMatchSet'

-> GeoMatchSet

Create a value of GeoMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GeoMatchSet', geoMatchSet_name - A friendly name or description of the GeoMatchSet. You can't change the name of an GeoMatchSet after you create it.

$sel:geoMatchSetId:GeoMatchSet', geoMatchSet_geoMatchSetId - The GeoMatchSetId for an GeoMatchSet. You use GeoMatchSetId to get information about a GeoMatchSet (see GeoMatchSet), update a GeoMatchSet (see UpdateGeoMatchSet), insert a GeoMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete a GeoMatchSet from AWS WAF (see DeleteGeoMatchSet).

GeoMatchSetId is returned by CreateGeoMatchSet and by ListGeoMatchSets.

$sel:geoMatchConstraints:GeoMatchSet', geoMatchSet_geoMatchConstraints - An array of GeoMatchConstraint objects, which contain the country that you want AWS WAF to search for.

geoMatchSet_name :: Lens' GeoMatchSet (Maybe Text)

A friendly name or description of the GeoMatchSet. You can't change the name of an GeoMatchSet after you create it.

geoMatchSet_geoMatchSetId :: Lens' GeoMatchSet Text

The GeoMatchSetId for an GeoMatchSet. You use GeoMatchSetId to get information about a GeoMatchSet (see GeoMatchSet), update a GeoMatchSet (see UpdateGeoMatchSet), insert a GeoMatchSet into a Rule or delete one from a Rule (see UpdateRule), and delete a GeoMatchSet from AWS WAF (see DeleteGeoMatchSet).

GeoMatchSetId is returned by CreateGeoMatchSet and by ListGeoMatchSets.

geoMatchSet_geoMatchConstraints :: Lens' GeoMatchSet [GeoMatchConstraint]

An array of GeoMatchConstraint objects, which contain the country that you want AWS WAF to search for.