Amazonka.WAFRegional.Types.RegexMatchSetUpdate

Description

Documentation

data RegexMatchSetUpdate

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

In an UpdateRegexMatchSet request, RegexMatchSetUpdate specifies whether to insert or delete a RegexMatchTuple and includes the settings for the RegexMatchTuple.

See: newRegexMatchSetUpdate smart constructor.

Constructors

RegexMatchSetUpdate'

Fields

action :: ChangeAction

Specifies whether to insert or delete a RegexMatchTuple.

regexMatchTuple :: RegexMatchTuple

Information about the part of a web request that you want AWS WAF to inspect and the identifier of the regular expression (regex) pattern that you want AWS WAF to search for. If you specify DELETE for the value of Action, the RegexMatchTuple values must exactly match the values in the RegexMatchTuple that you want to delete from the RegexMatchSet.

newRegexMatchSetUpdate

Arguments

:: ChangeAction
$sel:action:RegexMatchSetUpdate'

-> RegexMatchTuple
$sel:regexMatchTuple:RegexMatchSetUpdate'

-> RegexMatchSetUpdate

Create a value of RegexMatchSetUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:RegexMatchSetUpdate', regexMatchSetUpdate_action - Specifies whether to insert or delete a RegexMatchTuple.

$sel:regexMatchTuple:RegexMatchSetUpdate', regexMatchSetUpdate_regexMatchTuple - Information about the part of a web request that you want AWS WAF to inspect and the identifier of the regular expression (regex) pattern that you want AWS WAF to search for. If you specify DELETE for the value of Action, the RegexMatchTuple values must exactly match the values in the RegexMatchTuple that you want to delete from the RegexMatchSet.

regexMatchSetUpdate_action :: Lens' RegexMatchSetUpdate ChangeAction

Specifies whether to insert or delete a RegexMatchTuple.

regexMatchSetUpdate_regexMatchTuple :: Lens' RegexMatchSetUpdate RegexMatchTuple

Information about the part of a web request that you want AWS WAF to inspect and the identifier of the regular expression (regex) pattern that you want AWS WAF to search for. If you specify DELETE for the value of Action, the RegexMatchTuple values must exactly match the values in the RegexMatchTuple that you want to delete from the RegexMatchSet.