Amazonka.WAFRegional.Types.RuleGroupUpdate

Description

Documentation

data RuleGroupUpdate

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Specifies an ActivatedRule and indicates whether you want to add it to a RuleGroup or delete it from a RuleGroup.

See: newRuleGroupUpdate smart constructor.

Constructors

RuleGroupUpdate'

Fields

action :: ChangeAction

Specify INSERT to add an ActivatedRule to a RuleGroup. Use DELETE to remove an ActivatedRule from a RuleGroup.

activatedRule :: ActivatedRule

The ActivatedRule object specifies a Rule that you want to insert or delete, the priority of the Rule in the WebACL, and the action that you want AWS WAF to take when a web request matches the Rule (ALLOW, BLOCK, or COUNT).

newRuleGroupUpdate

Arguments

:: ChangeAction
$sel:action:RuleGroupUpdate'

-> ActivatedRule
$sel:activatedRule:RuleGroupUpdate'

-> RuleGroupUpdate

Create a value of RuleGroupUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:RuleGroupUpdate', ruleGroupUpdate_action - Specify INSERT to add an ActivatedRule to a RuleGroup. Use DELETE to remove an ActivatedRule from a RuleGroup.

$sel:activatedRule:RuleGroupUpdate', ruleGroupUpdate_activatedRule - The ActivatedRule object specifies a Rule that you want to insert or delete, the priority of the Rule in the WebACL, and the action that you want AWS WAF to take when a web request matches the Rule (ALLOW, BLOCK, or COUNT).

ruleGroupUpdate_action :: Lens' RuleGroupUpdate ChangeAction

Specify INSERT to add an ActivatedRule to a RuleGroup. Use DELETE to remove an ActivatedRule from a RuleGroup.

ruleGroupUpdate_activatedRule :: Lens' RuleGroupUpdate ActivatedRule

The ActivatedRule object specifies a Rule that you want to insert or delete, the priority of the Rule in the WebACL, and the action that you want AWS WAF to take when a web request matches the Rule (ALLOW, BLOCK, or COUNT).