Amazonka.WAFRegional.Types.WebACL

Description

Documentation

data WebACL

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Contains the Rules that identify the requests that you want to allow, block, or count. In a WebACL, you also specify a default action (ALLOW or BLOCK), and the action for each Rule that you add to a WebACL, for example, block requests from specified IP addresses or block requests from specified referrers. You also associate the WebACL with a CloudFront distribution to identify the requests that you want AWS WAF to filter. If you add more than one Rule to a WebACL, a request needs to match only one of the specifications to be allowed, blocked, or counted. For more information, see UpdateWebACL.

See: newWebACL smart constructor.

Constructors

WebACL'

Fields

metricName :: Maybe Text

A friendly name or description for the metrics for this WebACL. The name can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum length 128 and minimum length one. It can't contain whitespace or metric names reserved for AWS WAF, including "All" and "Default_Action." You can't change MetricName after you create the WebACL.

name :: Maybe Text

A friendly name or description of the WebACL. You can't change the name of a WebACL after you create it.

webACLArn :: Maybe Text

Tha Amazon Resource Name (ARN) of the web ACL.

webACLId :: Text

A unique identifier for a WebACL. You use WebACLId to get information about a WebACL (see GetWebACL), update a WebACL (see UpdateWebACL), and delete a WebACL from AWS WAF (see DeleteWebACL).

WebACLId is returned by CreateWebACL and by ListWebACLs.

defaultAction :: WafAction

The action to perform if none of the Rules contained in the WebACL match. The action is specified by the WafAction object.

rules :: [ActivatedRule]

An array that contains the action for each Rule in a WebACL, the priority of the Rule, and the ID of the Rule.

newWebACL

Arguments

:: Text
$sel:webACLId:WebACL'

-> WafAction
$sel:defaultAction:WebACL'

-> WebACL

Create a value of WebACL with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:WebACL', webACL_metricName - A friendly name or description for the metrics for this WebACL. The name can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum length 128 and minimum length one. It can't contain whitespace or metric names reserved for AWS WAF, including "All" and "Default_Action." You can't change MetricName after you create the WebACL.

$sel:name:WebACL', webACL_name - A friendly name or description of the WebACL. You can't change the name of a WebACL after you create it.

$sel:webACLArn:WebACL', webACL_webACLArn - Tha Amazon Resource Name (ARN) of the web ACL.

$sel:webACLId:WebACL', webACL_webACLId - A unique identifier for a WebACL. You use WebACLId to get information about a WebACL (see GetWebACL), update a WebACL (see UpdateWebACL), and delete a WebACL from AWS WAF (see DeleteWebACL).

WebACLId is returned by CreateWebACL and by ListWebACLs.

$sel:defaultAction:WebACL', webACL_defaultAction - The action to perform if none of the Rules contained in the WebACL match. The action is specified by the WafAction object.

$sel:rules:WebACL', webACL_rules - An array that contains the action for each Rule in a WebACL, the priority of the Rule, and the ID of the Rule.

webACL_metricName :: Lens' WebACL (Maybe Text)

A friendly name or description for the metrics for this WebACL. The name can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum length 128 and minimum length one. It can't contain whitespace or metric names reserved for AWS WAF, including "All" and "Default_Action." You can't change MetricName after you create the WebACL.

webACL_name :: Lens' WebACL (Maybe Text)

A friendly name or description of the WebACL. You can't change the name of a WebACL after you create it.

webACL_webACLArn :: Lens' WebACL (Maybe Text)

Tha Amazon Resource Name (ARN) of the web ACL.

webACL_webACLId :: Lens' WebACL Text

A unique identifier for a WebACL. You use WebACLId to get information about a WebACL (see GetWebACL), update a WebACL (see UpdateWebACL), and delete a WebACL from AWS WAF (see DeleteWebACL).

WebACLId is returned by CreateWebACL and by ListWebACLs.

webACL_defaultAction :: Lens' WebACL WafAction

The action to perform if none of the Rules contained in the WebACL match. The action is specified by the WafAction object.

webACL_rules :: Lens' WebACL [ActivatedRule]

An array that contains the action for each Rule in a WebACL, the priority of the Rule, and the ID of the Rule.