Amazonka.WAFRegional.UpdateSqlInjectionMatchSet

Description

This is AWS WAF Classic documentation. For more information, see AWS WAF Classic in the developer guide.

For the latest version of AWS WAF, use the AWS WAFV2 API and see the AWS WAF Developer Guide. With the latest version, AWS WAF has a single set of endpoints for regional and global use.

Inserts or deletes SqlInjectionMatchTuple objects (filters) in a SqlInjectionMatchSet. For each SqlInjectionMatchTuple object, you specify the following values:

Action: Whether to insert the object into or delete the object from the array. To change a SqlInjectionMatchTuple, you delete the existing object and add a new one.

FieldToMatch: The part of web requests that you want AWS WAF to inspect and, if you want AWS WAF to inspect a header or custom query parameter, the name of the header or parameter.

TextTransformation: Which text transformation, if any, to perform on the web request before inspecting the request for snippets of malicious SQL code.

You can only specify a single type of TextTransformation.

You use SqlInjectionMatchSet objects to specify which CloudFront requests that you want to allow, block, or count. For example, if you're receiving requests that contain snippets of SQL code in the query string and you want to block the requests, you can create a SqlInjectionMatchSet with the applicable settings, and then configure AWS WAF to block the requests.

To create and configure a SqlInjectionMatchSet, perform the following steps:

Submit a CreateSqlInjectionMatchSet request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateIPSet request.

Submit an UpdateSqlInjectionMatchSet request to specify the parts of web requests that you want AWS WAF to inspect for snippets of SQL code.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide.

Creating a Request

data UpdateSqlInjectionMatchSet

A request to update a SqlInjectionMatchSet.

See: newUpdateSqlInjectionMatchSet smart constructor.

Constructors

UpdateSqlInjectionMatchSet'

Fields

sqlInjectionMatchSetId :: Text

The SqlInjectionMatchSetId of the SqlInjectionMatchSet that you want to update. SqlInjectionMatchSetId is returned by CreateSqlInjectionMatchSet and by ListSqlInjectionMatchSets.

changeToken :: Text

The value returned by the most recent call to GetChangeToken.

updates :: NonEmpty SqlInjectionMatchSetUpdate

An array of SqlInjectionMatchSetUpdate objects that you want to insert into or delete from a SqlInjectionMatchSet. For more information, see the applicable data types:

SqlInjectionMatchSetUpdate: Contains Action and SqlInjectionMatchTuple

SqlInjectionMatchTuple: Contains FieldToMatch and TextTransformation

FieldToMatch: Contains Data and Type

newUpdateSqlInjectionMatchSet

Arguments

:: Text
$sel:sqlInjectionMatchSetId:UpdateSqlInjectionMatchSet'

-> Text
$sel:changeToken:UpdateSqlInjectionMatchSet'

-> NonEmpty SqlInjectionMatchSetUpdate
$sel:updates:UpdateSqlInjectionMatchSet'

-> UpdateSqlInjectionMatchSet

Create a value of UpdateSqlInjectionMatchSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sqlInjectionMatchSetId:UpdateSqlInjectionMatchSet', updateSqlInjectionMatchSet_sqlInjectionMatchSetId - The SqlInjectionMatchSetId of the SqlInjectionMatchSet that you want to update. SqlInjectionMatchSetId is returned by CreateSqlInjectionMatchSet and by ListSqlInjectionMatchSets.

$sel:changeToken:UpdateSqlInjectionMatchSet', updateSqlInjectionMatchSet_changeToken - The value returned by the most recent call to GetChangeToken.

$sel:updates:UpdateSqlInjectionMatchSet', updateSqlInjectionMatchSet_updates - An array of SqlInjectionMatchSetUpdate objects that you want to insert into or delete from a SqlInjectionMatchSet. For more information, see the applicable data types:

SqlInjectionMatchSetUpdate: Contains Action and SqlInjectionMatchTuple

SqlInjectionMatchTuple: Contains FieldToMatch and TextTransformation

FieldToMatch: Contains Data and Type

Request Lenses

updateSqlInjectionMatchSet_sqlInjectionMatchSetId :: Lens' UpdateSqlInjectionMatchSet Text

The SqlInjectionMatchSetId of the SqlInjectionMatchSet that you want to update. SqlInjectionMatchSetId is returned by CreateSqlInjectionMatchSet and by ListSqlInjectionMatchSets.

updateSqlInjectionMatchSet_changeToken :: Lens' UpdateSqlInjectionMatchSet Text

The value returned by the most recent call to GetChangeToken.

updateSqlInjectionMatchSet_updates :: Lens' UpdateSqlInjectionMatchSet (NonEmpty SqlInjectionMatchSetUpdate)

An array of SqlInjectionMatchSetUpdate objects that you want to insert into or delete from a SqlInjectionMatchSet. For more information, see the applicable data types:

SqlInjectionMatchSetUpdate: Contains Action and SqlInjectionMatchTuple

SqlInjectionMatchTuple: Contains FieldToMatch and TextTransformation

FieldToMatch: Contains Data and Type

Destructuring the Response

data UpdateSqlInjectionMatchSetResponse

The response to an UpdateSqlInjectionMatchSets request.

See: newUpdateSqlInjectionMatchSetResponse smart constructor.

Constructors

UpdateSqlInjectionMatchSetResponse'

Fields

changeToken :: Maybe Text

The ChangeToken that you used to submit the UpdateSqlInjectionMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

httpStatus :: Int

The response's http status code.

newUpdateSqlInjectionMatchSetResponse

Arguments

:: Int
$sel:httpStatus:UpdateSqlInjectionMatchSetResponse'

-> UpdateSqlInjectionMatchSetResponse

Create a value of UpdateSqlInjectionMatchSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:changeToken:UpdateSqlInjectionMatchSet', updateSqlInjectionMatchSetResponse_changeToken - The ChangeToken that you used to submit the UpdateSqlInjectionMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

$sel:httpStatus:UpdateSqlInjectionMatchSetResponse', updateSqlInjectionMatchSetResponse_httpStatus - The response's http status code.

Response Lenses

updateSqlInjectionMatchSetResponse_changeToken :: Lens' UpdateSqlInjectionMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the UpdateSqlInjectionMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus.

updateSqlInjectionMatchSetResponse_httpStatus :: Lens' UpdateSqlInjectionMatchSetResponse Int

The response's http status code.