Amazonka.WellArchitected.CreateWorkloadShare

Description

Create a workload share.

The owner of a workload can share it with other AWS accounts and IAM users in the same AWS Region. Shared access to a workload is not removed until the workload invitation is deleted.

For more information, see Sharing a Workload in the AWS Well-Architected Tool User Guide.

Creating a Request

data CreateWorkloadShare

Input for Create Workload Share

See: newCreateWorkloadShare smart constructor.

Constructors

CreateWorkloadShare'

Fields

workloadId :: Text

sharedWith :: Text

permissionType :: PermissionType

clientRequestToken :: Text

newCreateWorkloadShare

Arguments

:: Text
$sel:workloadId:CreateWorkloadShare'

-> Text
$sel:sharedWith:CreateWorkloadShare'

-> PermissionType
$sel:permissionType:CreateWorkloadShare'

-> Text
$sel:clientRequestToken:CreateWorkloadShare'

-> CreateWorkloadShare

Create a value of CreateWorkloadShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workloadId:CreateWorkloadShare', createWorkloadShare_workloadId - Undocumented member.

$sel:sharedWith:CreateWorkloadShare', createWorkloadShare_sharedWith - Undocumented member.

$sel:permissionType:CreateWorkloadShare', createWorkloadShare_permissionType - Undocumented member.

$sel:clientRequestToken:CreateWorkloadShare', createWorkloadShare_clientRequestToken - Undocumented member.

Request Lenses

createWorkloadShare_workloadId :: Lens' CreateWorkloadShare Text

Undocumented member.

createWorkloadShare_sharedWith :: Lens' CreateWorkloadShare Text

Undocumented member.

createWorkloadShare_permissionType :: Lens' CreateWorkloadShare PermissionType

Undocumented member.

createWorkloadShare_clientRequestToken :: Lens' CreateWorkloadShare Text

Undocumented member.

Destructuring the Response

data CreateWorkloadShareResponse

Input for Create Workload Share

See: newCreateWorkloadShareResponse smart constructor.

Constructors

CreateWorkloadShareResponse'

Fields

workloadId :: Maybe Text

shareId :: Maybe Text

httpStatus :: Int

The response's http status code.

newCreateWorkloadShareResponse

Arguments

:: Int
$sel:httpStatus:CreateWorkloadShareResponse'

-> CreateWorkloadShareResponse

Create a value of CreateWorkloadShareResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workloadId:CreateWorkloadShare', createWorkloadShareResponse_workloadId - Undocumented member.

$sel:shareId:CreateWorkloadShareResponse', createWorkloadShareResponse_shareId - Undocumented member.

$sel:httpStatus:CreateWorkloadShareResponse', createWorkloadShareResponse_httpStatus - The response's http status code.

Response Lenses

createWorkloadShareResponse_workloadId :: Lens' CreateWorkloadShareResponse (Maybe Text)

Undocumented member.

createWorkloadShareResponse_shareId :: Lens' CreateWorkloadShareResponse (Maybe Text)

Undocumented member.

createWorkloadShareResponse_httpStatus :: Lens' CreateWorkloadShareResponse Int

The response's http status code.