Amazonka.WellArchitected.GetAnswer

Description

Get the answer to a specific question in a workload review.

Creating a Request

data GetAnswer

Input to get answer.

See: newGetAnswer smart constructor.

Constructors

GetAnswer'

Fields

milestoneNumber :: Maybe Natural

workloadId :: Text

lensAlias :: Text

questionId :: Text

newGetAnswer

Arguments

:: Text
$sel:workloadId:GetAnswer'

-> Text
$sel:lensAlias:GetAnswer'

-> Text
$sel:questionId:GetAnswer'

-> GetAnswer

Create a value of GetAnswer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:milestoneNumber:GetAnswer', getAnswer_milestoneNumber - Undocumented member.

$sel:workloadId:GetAnswer', getAnswer_workloadId - Undocumented member.

$sel:lensAlias:GetAnswer', getAnswer_lensAlias - Undocumented member.

$sel:questionId:GetAnswer', getAnswer_questionId - Undocumented member.

Request Lenses

getAnswer_milestoneNumber :: Lens' GetAnswer (Maybe Natural)

Undocumented member.

getAnswer_workloadId :: Lens' GetAnswer Text

Undocumented member.

getAnswer_lensAlias :: Lens' GetAnswer Text

Undocumented member.

getAnswer_questionId :: Lens' GetAnswer Text

Undocumented member.

Destructuring the Response

data GetAnswerResponse

Output of a get answer call.

See: newGetAnswerResponse smart constructor.

Constructors

GetAnswerResponse'

Fields

lensAlias :: Maybe Text

milestoneNumber :: Maybe Natural

answer :: Maybe Answer

workloadId :: Maybe Text

httpStatus :: Int

The response's http status code.

newGetAnswerResponse

Arguments

:: Int
$sel:httpStatus:GetAnswerResponse'

-> GetAnswerResponse

Create a value of GetAnswerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lensAlias:GetAnswer', getAnswerResponse_lensAlias - Undocumented member.

$sel:milestoneNumber:GetAnswer', getAnswerResponse_milestoneNumber - Undocumented member.

$sel:answer:GetAnswerResponse', getAnswerResponse_answer - Undocumented member.

$sel:workloadId:GetAnswer', getAnswerResponse_workloadId - Undocumented member.

$sel:httpStatus:GetAnswerResponse', getAnswerResponse_httpStatus - The response's http status code.

Response Lenses

getAnswerResponse_lensAlias :: Lens' GetAnswerResponse (Maybe Text)

Undocumented member.

getAnswerResponse_milestoneNumber :: Lens' GetAnswerResponse (Maybe Natural)

Undocumented member.

getAnswerResponse_answer :: Lens' GetAnswerResponse (Maybe Answer)

Undocumented member.

getAnswerResponse_workloadId :: Lens' GetAnswerResponse (Maybe Text)

Undocumented member.

getAnswerResponse_httpStatus :: Lens' GetAnswerResponse Int

The response's http status code.