Amazonka.WellArchitected.Lens

Description

Operations

ListNotifications

listNotifications_nextToken :: Lens' ListNotifications (Maybe Text)

Undocumented member.

listNotifications_workloadId :: Lens' ListNotifications (Maybe Text)

Undocumented member.

listNotifications_maxResults :: Lens' ListNotifications (Maybe Natural)

The maximum number of results to return for this request.

listNotificationsResponse_notificationSummaries :: Lens' ListNotificationsResponse (Maybe [NotificationSummary])

List of lens notification summaries in a workload.

listNotificationsResponse_nextToken :: Lens' ListNotificationsResponse (Maybe Text)

Undocumented member.

listNotificationsResponse_httpStatus :: Lens' ListNotificationsResponse Int

The response's http status code.

GetLensVersionDifference

getLensVersionDifference_lensAlias :: Lens' GetLensVersionDifference Text

Undocumented member.

getLensVersionDifference_baseLensVersion :: Lens' GetLensVersionDifference Text

The base version of the lens.

getLensVersionDifferenceResponse_lensAlias :: Lens' GetLensVersionDifferenceResponse (Maybe Text)

Undocumented member.

getLensVersionDifferenceResponse_latestLensVersion :: Lens' GetLensVersionDifferenceResponse (Maybe Text)

The latest version of the lens.

getLensVersionDifferenceResponse_baseLensVersion :: Lens' GetLensVersionDifferenceResponse (Maybe Text)

The base version of the lens.

getLensVersionDifferenceResponse_versionDifferences :: Lens' GetLensVersionDifferenceResponse (Maybe VersionDifferences)

Undocumented member.

getLensVersionDifferenceResponse_httpStatus :: Lens' GetLensVersionDifferenceResponse Int

The response's http status code.

ListLensReviewImprovements

listLensReviewImprovements_pillarId :: Lens' ListLensReviewImprovements (Maybe Text)

Undocumented member.

listLensReviewImprovements_milestoneNumber :: Lens' ListLensReviewImprovements (Maybe Natural)

Undocumented member.

listLensReviewImprovements_nextToken :: Lens' ListLensReviewImprovements (Maybe Text)

Undocumented member.

listLensReviewImprovements_maxResults :: Lens' ListLensReviewImprovements (Maybe Natural)

The maximum number of results to return for this request.

listLensReviewImprovements_workloadId :: Lens' ListLensReviewImprovements Text

Undocumented member.

listLensReviewImprovements_lensAlias :: Lens' ListLensReviewImprovements Text

Undocumented member.

listLensReviewImprovementsResponse_improvementSummaries :: Lens' ListLensReviewImprovementsResponse (Maybe [ImprovementSummary])

Undocumented member.

listLensReviewImprovementsResponse_lensAlias :: Lens' ListLensReviewImprovementsResponse (Maybe Text)

Undocumented member.

listLensReviewImprovementsResponse_milestoneNumber :: Lens' ListLensReviewImprovementsResponse (Maybe Natural)

Undocumented member.

listLensReviewImprovementsResponse_nextToken :: Lens' ListLensReviewImprovementsResponse (Maybe Text)

Undocumented member.

listLensReviewImprovementsResponse_workloadId :: Lens' ListLensReviewImprovementsResponse (Maybe Text)

Undocumented member.

listLensReviewImprovementsResponse_httpStatus :: Lens' ListLensReviewImprovementsResponse Int

The response's http status code.

ListMilestones

listMilestones_nextToken :: Lens' ListMilestones (Maybe Text)

Undocumented member.

listMilestones_maxResults :: Lens' ListMilestones (Maybe Natural)

Undocumented member.

listMilestones_workloadId :: Lens' ListMilestones Text

Undocumented member.

listMilestonesResponse_milestoneSummaries :: Lens' ListMilestonesResponse (Maybe [MilestoneSummary])

Undocumented member.

listMilestonesResponse_nextToken :: Lens' ListMilestonesResponse (Maybe Text)

Undocumented member.

listMilestonesResponse_workloadId :: Lens' ListMilestonesResponse (Maybe Text)

Undocumented member.

listMilestonesResponse_httpStatus :: Lens' ListMilestonesResponse Int

The response's http status code.

CreateMilestone

createMilestone_workloadId :: Lens' CreateMilestone Text

Undocumented member.

createMilestone_milestoneName :: Lens' CreateMilestone Text

Undocumented member.

createMilestone_clientRequestToken :: Lens' CreateMilestone Text

Undocumented member.

createMilestoneResponse_milestoneNumber :: Lens' CreateMilestoneResponse (Maybe Natural)

Undocumented member.

createMilestoneResponse_workloadId :: Lens' CreateMilestoneResponse (Maybe Text)

Undocumented member.

createMilestoneResponse_httpStatus :: Lens' CreateMilestoneResponse Int

The response's http status code.

GetAnswer

getAnswer_milestoneNumber :: Lens' GetAnswer (Maybe Natural)

Undocumented member.

getAnswer_workloadId :: Lens' GetAnswer Text

Undocumented member.

getAnswer_lensAlias :: Lens' GetAnswer Text

Undocumented member.

getAnswer_questionId :: Lens' GetAnswer Text

Undocumented member.

getAnswerResponse_lensAlias :: Lens' GetAnswerResponse (Maybe Text)

Undocumented member.

getAnswerResponse_milestoneNumber :: Lens' GetAnswerResponse (Maybe Natural)

Undocumented member.

getAnswerResponse_answer :: Lens' GetAnswerResponse (Maybe Answer)

Undocumented member.

getAnswerResponse_workloadId :: Lens' GetAnswerResponse (Maybe Text)

Undocumented member.

getAnswerResponse_httpStatus :: Lens' GetAnswerResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_workloadArn :: Lens' ListTagsForResource Text

Undocumented member.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags for the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

UpdateAnswer

updateAnswer_isApplicable :: Lens' UpdateAnswer (Maybe Bool)

Undocumented member.

updateAnswer_selectedChoices :: Lens' UpdateAnswer (Maybe [Text])

Undocumented member.

updateAnswer_reason :: Lens' UpdateAnswer (Maybe AnswerReason)

The reason why a question is not applicable to your workload.

updateAnswer_notes :: Lens' UpdateAnswer (Maybe Text)

Undocumented member.

updateAnswer_choiceUpdates :: Lens' UpdateAnswer (Maybe (HashMap Text ChoiceUpdate))

A list of choices to update on a question in your workload. The String key corresponds to the choice ID to be updated.

updateAnswer_workloadId :: Lens' UpdateAnswer Text

Undocumented member.

updateAnswer_lensAlias :: Lens' UpdateAnswer Text

Undocumented member.

updateAnswer_questionId :: Lens' UpdateAnswer Text

Undocumented member.

updateAnswerResponse_lensAlias :: Lens' UpdateAnswerResponse (Maybe Text)

Undocumented member.

updateAnswerResponse_answer :: Lens' UpdateAnswerResponse (Maybe Answer)

Undocumented member.

updateAnswerResponse_workloadId :: Lens' UpdateAnswerResponse (Maybe Text)

Undocumented member.

updateAnswerResponse_httpStatus :: Lens' UpdateAnswerResponse Int

The response's http status code.

UpdateShareInvitation

updateShareInvitation_shareInvitationId :: Lens' UpdateShareInvitation Text

The ID assigned to the share invitation.

updateShareInvitation_shareInvitationAction :: Lens' UpdateShareInvitation ShareInvitationAction

Undocumented member.

updateShareInvitationResponse_shareInvitation :: Lens' UpdateShareInvitationResponse (Maybe ShareInvitation)

The updated workload share invitation.

updateShareInvitationResponse_httpStatus :: Lens' UpdateShareInvitationResponse Int

The response's http status code.

ListAnswers

listAnswers_pillarId :: Lens' ListAnswers (Maybe Text)

Undocumented member.

listAnswers_milestoneNumber :: Lens' ListAnswers (Maybe Natural)

Undocumented member.

listAnswers_nextToken :: Lens' ListAnswers (Maybe Text)

Undocumented member.

listAnswers_maxResults :: Lens' ListAnswers (Maybe Natural)

The maximum number of results to return for this request.

listAnswers_workloadId :: Lens' ListAnswers Text

Undocumented member.

listAnswers_lensAlias :: Lens' ListAnswers Text

Undocumented member.

listAnswersResponse_lensAlias :: Lens' ListAnswersResponse (Maybe Text)

Undocumented member.

listAnswersResponse_milestoneNumber :: Lens' ListAnswersResponse (Maybe Natural)

Undocumented member.

listAnswersResponse_nextToken :: Lens' ListAnswersResponse (Maybe Text)

Undocumented member.

listAnswersResponse_workloadId :: Lens' ListAnswersResponse (Maybe Text)

Undocumented member.

listAnswersResponse_answerSummaries :: Lens' ListAnswersResponse (Maybe [AnswerSummary])

Undocumented member.

listAnswersResponse_httpStatus :: Lens' ListAnswersResponse Int

The response's http status code.

DisassociateLenses

disassociateLenses_workloadId :: Lens' DisassociateLenses Text

Undocumented member.

disassociateLenses_lensAliases :: Lens' DisassociateLenses (NonEmpty Text)

Undocumented member.

GetMilestone

getMilestone_workloadId :: Lens' GetMilestone Text

Undocumented member.

getMilestone_milestoneNumber :: Lens' GetMilestone Natural

Undocumented member.

getMilestoneResponse_milestone :: Lens' GetMilestoneResponse (Maybe Milestone)

Undocumented member.

getMilestoneResponse_workloadId :: Lens' GetMilestoneResponse (Maybe Text)

Undocumented member.

getMilestoneResponse_httpStatus :: Lens' GetMilestoneResponse Int

The response's http status code.

ListLenses

listLenses_nextToken :: Lens' ListLenses (Maybe Text)

Undocumented member.

listLenses_maxResults :: Lens' ListLenses (Maybe Natural)

Undocumented member.

listLensesResponse_nextToken :: Lens' ListLensesResponse (Maybe Text)

Undocumented member.

listLensesResponse_lensSummaries :: Lens' ListLensesResponse (Maybe [LensSummary])

Undocumented member.

listLensesResponse_httpStatus :: Lens' ListLensesResponse Int

The response's http status code.

ListWorkloadShares

listWorkloadShares_sharedWithPrefix :: Lens' ListWorkloadShares (Maybe Text)

The AWS account ID or IAM role with which the workload is shared.

listWorkloadShares_nextToken :: Lens' ListWorkloadShares (Maybe Text)

Undocumented member.

listWorkloadShares_maxResults :: Lens' ListWorkloadShares (Maybe Natural)

The maximum number of results to return for this request.

listWorkloadShares_workloadId :: Lens' ListWorkloadShares Text

Undocumented member.

listWorkloadSharesResponse_workloadShareSummaries :: Lens' ListWorkloadSharesResponse (Maybe [WorkloadShareSummary])

Undocumented member.

listWorkloadSharesResponse_nextToken :: Lens' ListWorkloadSharesResponse (Maybe Text)

Undocumented member.

listWorkloadSharesResponse_workloadId :: Lens' ListWorkloadSharesResponse (Maybe Text)

Undocumented member.

listWorkloadSharesResponse_httpStatus :: Lens' ListWorkloadSharesResponse Int

The response's http status code.

UpdateWorkload

updateWorkload_isReviewOwnerUpdateAcknowledged :: Lens' UpdateWorkload (Maybe Bool)

Flag indicating whether the workload owner has acknowledged that the Review owner field is required.

If a Review owner is not added to the workload within 60 days of acknowledgement, access to the workload is restricted until an owner is added.

updateWorkload_architecturalDesign :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_accountIds :: Lens' UpdateWorkload (Maybe [Text])

Undocumented member.

updateWorkload_industry :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_environment :: Lens' UpdateWorkload (Maybe WorkloadEnvironment)

Undocumented member.

updateWorkload_awsRegions :: Lens' UpdateWorkload (Maybe [Text])

Undocumented member.

updateWorkload_improvementStatus :: Lens' UpdateWorkload (Maybe WorkloadImprovementStatus)

Undocumented member.

updateWorkload_industryType :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_workloadName :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_notes :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_reviewOwner :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_description :: Lens' UpdateWorkload (Maybe Text)

Undocumented member.

updateWorkload_pillarPriorities :: Lens' UpdateWorkload (Maybe [Text])

Undocumented member.

updateWorkload_nonAwsRegions :: Lens' UpdateWorkload (Maybe [Text])

Undocumented member.

updateWorkload_workloadId :: Lens' UpdateWorkload Text

Undocumented member.

updateWorkloadResponse_workload :: Lens' UpdateWorkloadResponse (Maybe Workload)

Undocumented member.

updateWorkloadResponse_httpStatus :: Lens' UpdateWorkloadResponse Int

The response's http status code.

DeleteWorkload

deleteWorkload_workloadId :: Lens' DeleteWorkload Text

Undocumented member.

deleteWorkload_clientRequestToken :: Lens' DeleteWorkload Text

Undocumented member.

ListLensReviews

listLensReviews_milestoneNumber :: Lens' ListLensReviews (Maybe Natural)

Undocumented member.

listLensReviews_nextToken :: Lens' ListLensReviews (Maybe Text)

Undocumented member.

listLensReviews_maxResults :: Lens' ListLensReviews (Maybe Natural)

Undocumented member.

listLensReviews_workloadId :: Lens' ListLensReviews Text

Undocumented member.

listLensReviewsResponse_milestoneNumber :: Lens' ListLensReviewsResponse (Maybe Natural)

Undocumented member.

listLensReviewsResponse_nextToken :: Lens' ListLensReviewsResponse (Maybe Text)

Undocumented member.

listLensReviewsResponse_lensReviewSummaries :: Lens' ListLensReviewsResponse (Maybe [LensReviewSummary])

Undocumented member.

listLensReviewsResponse_workloadId :: Lens' ListLensReviewsResponse (Maybe Text)

Undocumented member.

listLensReviewsResponse_httpStatus :: Lens' ListLensReviewsResponse Int

The response's http status code.

UpdateLensReview

updateLensReview_lensNotes :: Lens' UpdateLensReview (Maybe Text)

Undocumented member.

updateLensReview_pillarNotes :: Lens' UpdateLensReview (Maybe (HashMap Text Text))

Undocumented member.

updateLensReview_workloadId :: Lens' UpdateLensReview Text

Undocumented member.

updateLensReview_lensAlias :: Lens' UpdateLensReview Text

Undocumented member.

updateLensReviewResponse_workloadId :: Lens' UpdateLensReviewResponse (Maybe Text)

Undocumented member.

updateLensReviewResponse_lensReview :: Lens' UpdateLensReviewResponse (Maybe LensReview)

Undocumented member.

updateLensReviewResponse_httpStatus :: Lens' UpdateLensReviewResponse Int

The response's http status code.

ListShareInvitations

listShareInvitations_nextToken :: Lens' ListShareInvitations (Maybe Text)

Undocumented member.

listShareInvitations_workloadNamePrefix :: Lens' ListShareInvitations (Maybe Text)

Undocumented member.

listShareInvitations_maxResults :: Lens' ListShareInvitations (Maybe Natural)

The maximum number of results to return for this request.

listShareInvitationsResponse_shareInvitationSummaries :: Lens' ListShareInvitationsResponse (Maybe [ShareInvitationSummary])

List of share invitation summaries in a workload.

listShareInvitationsResponse_nextToken :: Lens' ListShareInvitationsResponse (Maybe Text)

Undocumented member.

listShareInvitationsResponse_httpStatus :: Lens' ListShareInvitationsResponse Int

The response's http status code.

GetLensReview

getLensReview_milestoneNumber :: Lens' GetLensReview (Maybe Natural)

Undocumented member.

getLensReview_workloadId :: Lens' GetLensReview Text

Undocumented member.

getLensReview_lensAlias :: Lens' GetLensReview Text

Undocumented member.

getLensReviewResponse_milestoneNumber :: Lens' GetLensReviewResponse (Maybe Natural)

Undocumented member.

getLensReviewResponse_workloadId :: Lens' GetLensReviewResponse (Maybe Text)

Undocumented member.

getLensReviewResponse_lensReview :: Lens' GetLensReviewResponse (Maybe LensReview)

Undocumented member.

getLensReviewResponse_httpStatus :: Lens' GetLensReviewResponse Int

The response's http status code.

TagResource

tagResource_workloadArn :: Lens' TagResource Text

Undocumented member.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags for the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

CreateWorkload

createWorkload_architecturalDesign :: Lens' CreateWorkload (Maybe Text)

Undocumented member.

createWorkload_accountIds :: Lens' CreateWorkload (Maybe [Text])

Undocumented member.

createWorkload_industry :: Lens' CreateWorkload (Maybe Text)

Undocumented member.

createWorkload_awsRegions :: Lens' CreateWorkload (Maybe [Text])

Undocumented member.

createWorkload_industryType :: Lens' CreateWorkload (Maybe Text)

Undocumented member.

createWorkload_notes :: Lens' CreateWorkload (Maybe Text)

Undocumented member.

createWorkload_pillarPriorities :: Lens' CreateWorkload (Maybe [Text])

Undocumented member.

createWorkload_nonAwsRegions :: Lens' CreateWorkload (Maybe [Text])

Undocumented member.

createWorkload_tags :: Lens' CreateWorkload (Maybe (HashMap Text Text))

The tags to be associated with the workload.

createWorkload_workloadName :: Lens' CreateWorkload Text

Undocumented member.

createWorkload_description :: Lens' CreateWorkload Text

Undocumented member.

createWorkload_environment :: Lens' CreateWorkload WorkloadEnvironment

Undocumented member.

createWorkload_reviewOwner :: Lens' CreateWorkload Text

Undocumented member.

createWorkload_lenses :: Lens' CreateWorkload [Text]

Undocumented member.

createWorkload_clientRequestToken :: Lens' CreateWorkload Text

Undocumented member.

createWorkloadResponse_workloadArn :: Lens' CreateWorkloadResponse (Maybe Text)

Undocumented member.

createWorkloadResponse_workloadId :: Lens' CreateWorkloadResponse (Maybe Text)

Undocumented member.

createWorkloadResponse_httpStatus :: Lens' CreateWorkloadResponse Int

The response's http status code.

DeleteWorkloadShare

deleteWorkloadShare_shareId :: Lens' DeleteWorkloadShare Text

Undocumented member.

deleteWorkloadShare_workloadId :: Lens' DeleteWorkloadShare Text

Undocumented member.

deleteWorkloadShare_clientRequestToken :: Lens' DeleteWorkloadShare Text

Undocumented member.

UpdateWorkloadShare

updateWorkloadShare_shareId :: Lens' UpdateWorkloadShare Text

Undocumented member.

updateWorkloadShare_workloadId :: Lens' UpdateWorkloadShare Text

Undocumented member.

updateWorkloadShare_permissionType :: Lens' UpdateWorkloadShare PermissionType

Undocumented member.

updateWorkloadShareResponse_workloadShare :: Lens' UpdateWorkloadShareResponse (Maybe WorkloadShare)

Undocumented member.

updateWorkloadShareResponse_workloadId :: Lens' UpdateWorkloadShareResponse (Maybe Text)

Undocumented member.

updateWorkloadShareResponse_httpStatus :: Lens' UpdateWorkloadShareResponse Int

The response's http status code.

UntagResource

untagResource_workloadArn :: Lens' UntagResource Text

Undocumented member.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

A list of tag keys. Existing tags of the resource whose keys are members of this list are removed from the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

AssociateLenses

associateLenses_workloadId :: Lens' AssociateLenses Text

Undocumented member.

associateLenses_lensAliases :: Lens' AssociateLenses (NonEmpty Text)

Undocumented member.

ListWorkloads

listWorkloads_nextToken :: Lens' ListWorkloads (Maybe Text)

Undocumented member.

listWorkloads_workloadNamePrefix :: Lens' ListWorkloads (Maybe Text)

Undocumented member.

listWorkloads_maxResults :: Lens' ListWorkloads (Maybe Natural)

The maximum number of results to return for this request.

listWorkloadsResponse_workloadSummaries :: Lens' ListWorkloadsResponse (Maybe [WorkloadSummary])

Undocumented member.

listWorkloadsResponse_nextToken :: Lens' ListWorkloadsResponse (Maybe Text)

Undocumented member.

listWorkloadsResponse_httpStatus :: Lens' ListWorkloadsResponse Int

The response's http status code.

CreateWorkloadShare

createWorkloadShare_workloadId :: Lens' CreateWorkloadShare Text

Undocumented member.

createWorkloadShare_sharedWith :: Lens' CreateWorkloadShare Text

Undocumented member.

createWorkloadShare_permissionType :: Lens' CreateWorkloadShare PermissionType

Undocumented member.

createWorkloadShare_clientRequestToken :: Lens' CreateWorkloadShare Text

Undocumented member.

createWorkloadShareResponse_workloadId :: Lens' CreateWorkloadShareResponse (Maybe Text)

Undocumented member.

createWorkloadShareResponse_shareId :: Lens' CreateWorkloadShareResponse (Maybe Text)

Undocumented member.

createWorkloadShareResponse_httpStatus :: Lens' CreateWorkloadShareResponse Int

The response's http status code.

GetLensReviewReport

getLensReviewReport_milestoneNumber :: Lens' GetLensReviewReport (Maybe Natural)

Undocumented member.

getLensReviewReport_workloadId :: Lens' GetLensReviewReport Text

Undocumented member.

getLensReviewReport_lensAlias :: Lens' GetLensReviewReport Text

Undocumented member.

getLensReviewReportResponse_milestoneNumber :: Lens' GetLensReviewReportResponse (Maybe Natural)

Undocumented member.

getLensReviewReportResponse_workloadId :: Lens' GetLensReviewReportResponse (Maybe Text)

Undocumented member.

getLensReviewReportResponse_lensReviewReport :: Lens' GetLensReviewReportResponse (Maybe LensReviewReport)

Undocumented member.

getLensReviewReportResponse_httpStatus :: Lens' GetLensReviewReportResponse Int

The response's http status code.

UpgradeLensReview

upgradeLensReview_clientRequestToken :: Lens' UpgradeLensReview (Maybe Text)

Undocumented member.

upgradeLensReview_workloadId :: Lens' UpgradeLensReview Text

Undocumented member.

upgradeLensReview_lensAlias :: Lens' UpgradeLensReview Text

Undocumented member.

upgradeLensReview_milestoneName :: Lens' UpgradeLensReview Text

Undocumented member.

GetWorkload

getWorkload_workloadId :: Lens' GetWorkload Text

Undocumented member.

getWorkloadResponse_workload :: Lens' GetWorkloadResponse (Maybe Workload)

Undocumented member.

getWorkloadResponse_httpStatus :: Lens' GetWorkloadResponse Int

The response's http status code.

Types

Answer

answer_choiceAnswers :: Lens' Answer (Maybe [ChoiceAnswer])

A list of selected choices to a question in your workload.

answer_helpfulResourceUrl :: Lens' Answer (Maybe Text)

Undocumented member.

answer_isApplicable :: Lens' Answer (Maybe Bool)

Undocumented member.

answer_pillarId :: Lens' Answer (Maybe Text)

Undocumented member.

answer_improvementPlanUrl :: Lens' Answer (Maybe Text)

Undocumented member.

answer_questionDescription :: Lens' Answer (Maybe Text)

Undocumented member.

answer_risk :: Lens' Answer (Maybe Risk)

Undocumented member.

answer_questionTitle :: Lens' Answer (Maybe Text)

Undocumented member.

answer_selectedChoices :: Lens' Answer (Maybe [Text])

Undocumented member.

answer_reason :: Lens' Answer (Maybe AnswerReason)

The reason why the question is not applicable to your workload.

answer_choices :: Lens' Answer (Maybe [Choice])

Undocumented member.

answer_questionId :: Lens' Answer (Maybe Text)

Undocumented member.

answer_notes :: Lens' Answer (Maybe Text)

Undocumented member.

AnswerSummary

answerSummary_isApplicable :: Lens' AnswerSummary (Maybe Bool)

Undocumented member.

answerSummary_pillarId :: Lens' AnswerSummary (Maybe Text)

Undocumented member.

answerSummary_choiceAnswerSummaries :: Lens' AnswerSummary (Maybe [ChoiceAnswerSummary])

A list of selected choices to a question in your workload.

answerSummary_risk :: Lens' AnswerSummary (Maybe Risk)

Undocumented member.

answerSummary_questionTitle :: Lens' AnswerSummary (Maybe Text)

Undocumented member.

answerSummary_selectedChoices :: Lens' AnswerSummary (Maybe [Text])

Undocumented member.

answerSummary_reason :: Lens' AnswerSummary (Maybe AnswerReason)

The reason why a choice is non-applicable to a question in your workload.

answerSummary_choices :: Lens' AnswerSummary (Maybe [Choice])

Undocumented member.

answerSummary_questionId :: Lens' AnswerSummary (Maybe Text)

Undocumented member.

Choice

choice_title :: Lens' Choice (Maybe Text)

Undocumented member.

choice_description :: Lens' Choice (Maybe Text)

Undocumented member.

choice_choiceId :: Lens' Choice (Maybe Text)

Undocumented member.

ChoiceAnswer

choiceAnswer_status :: Lens' ChoiceAnswer (Maybe ChoiceStatus)

The status of a choice.

choiceAnswer_reason :: Lens' ChoiceAnswer (Maybe ChoiceReason)

The reason why a choice is non-applicable to a question in your workload.

choiceAnswer_notes :: Lens' ChoiceAnswer (Maybe Text)

The notes associated with a choice.

choiceAnswer_choiceId :: Lens' ChoiceAnswer (Maybe Text)

Undocumented member.

ChoiceAnswerSummary

choiceAnswerSummary_status :: Lens' ChoiceAnswerSummary (Maybe ChoiceStatus)

The status of a choice.

choiceAnswerSummary_reason :: Lens' ChoiceAnswerSummary (Maybe ChoiceReason)

The reason why a choice is non-applicable to a question in your workload.

choiceAnswerSummary_choiceId :: Lens' ChoiceAnswerSummary (Maybe Text)

Undocumented member.

ChoiceUpdate

choiceUpdate_reason :: Lens' ChoiceUpdate (Maybe ChoiceReason)

The reason why a choice is non-applicable to a question in your workload.

choiceUpdate_notes :: Lens' ChoiceUpdate (Maybe Text)

The notes associated with a choice.

choiceUpdate_status :: Lens' ChoiceUpdate ChoiceStatus

The status of a choice.

ImprovementSummary

improvementSummary_pillarId :: Lens' ImprovementSummary (Maybe Text)

Undocumented member.

improvementSummary_improvementPlanUrl :: Lens' ImprovementSummary (Maybe Text)

Undocumented member.

improvementSummary_risk :: Lens' ImprovementSummary (Maybe Risk)

Undocumented member.

improvementSummary_questionTitle :: Lens' ImprovementSummary (Maybe Text)

Undocumented member.

improvementSummary_questionId :: Lens' ImprovementSummary (Maybe Text)

Undocumented member.

LensReview

lensReview_lensAlias :: Lens' LensReview (Maybe Text)

Undocumented member.

lensReview_riskCounts :: Lens' LensReview (Maybe (HashMap Risk Natural))

Undocumented member.

lensReview_lensName :: Lens' LensReview (Maybe Text)

Undocumented member.

lensReview_nextToken :: Lens' LensReview (Maybe Text)

Undocumented member.

lensReview_pillarReviewSummaries :: Lens' LensReview (Maybe [PillarReviewSummary])

Undocumented member.

lensReview_updatedAt :: Lens' LensReview (Maybe UTCTime)

Undocumented member.

lensReview_lensStatus :: Lens' LensReview (Maybe LensStatus)

The status of the lens.

lensReview_notes :: Lens' LensReview (Maybe Text)

Undocumented member.

lensReview_lensVersion :: Lens' LensReview (Maybe Text)

The version of the lens.

LensReviewReport

lensReviewReport_lensAlias :: Lens' LensReviewReport (Maybe Text)

Undocumented member.

lensReviewReport_base64String :: Lens' LensReviewReport (Maybe Text)

Undocumented member.

LensReviewSummary

lensReviewSummary_lensAlias :: Lens' LensReviewSummary (Maybe Text)

Undocumented member.

lensReviewSummary_riskCounts :: Lens' LensReviewSummary (Maybe (HashMap Risk Natural))

Undocumented member.

lensReviewSummary_lensName :: Lens' LensReviewSummary (Maybe Text)

Undocumented member.

lensReviewSummary_updatedAt :: Lens' LensReviewSummary (Maybe UTCTime)

Undocumented member.

lensReviewSummary_lensStatus :: Lens' LensReviewSummary (Maybe LensStatus)

The status of the lens.

lensReviewSummary_lensVersion :: Lens' LensReviewSummary (Maybe Text)

The version of the lens.

LensSummary

lensSummary_lensAlias :: Lens' LensSummary (Maybe Text)

Undocumented member.

lensSummary_lensName :: Lens' LensSummary (Maybe Text)

Undocumented member.

lensSummary_lensVersion :: Lens' LensSummary (Maybe Text)

The version of the lens.

lensSummary_description :: Lens' LensSummary (Maybe Text)

Undocumented member.

LensUpgradeSummary

lensUpgradeSummary_lensAlias :: Lens' LensUpgradeSummary (Maybe Text)

Undocumented member.

lensUpgradeSummary_latestLensVersion :: Lens' LensUpgradeSummary (Maybe Text)

The latest version of the lens.

lensUpgradeSummary_currentLensVersion :: Lens' LensUpgradeSummary (Maybe Text)

The current version of the lens.

lensUpgradeSummary_workloadId :: Lens' LensUpgradeSummary (Maybe Text)

Undocumented member.

lensUpgradeSummary_workloadName :: Lens' LensUpgradeSummary (Maybe Text)

Undocumented member.

Milestone

milestone_workload :: Lens' Milestone (Maybe Workload)

Undocumented member.

milestone_milestoneNumber :: Lens' Milestone (Maybe Natural)

Undocumented member.

milestone_milestoneName :: Lens' Milestone (Maybe Text)

Undocumented member.

milestone_recordedAt :: Lens' Milestone (Maybe UTCTime)

Undocumented member.

MilestoneSummary

milestoneSummary_milestoneNumber :: Lens' MilestoneSummary (Maybe Natural)

Undocumented member.

milestoneSummary_milestoneName :: Lens' MilestoneSummary (Maybe Text)

Undocumented member.

milestoneSummary_recordedAt :: Lens' MilestoneSummary (Maybe UTCTime)

Undocumented member.

milestoneSummary_workloadSummary :: Lens' MilestoneSummary (Maybe WorkloadSummary)

Undocumented member.

NotificationSummary

notificationSummary_lensUpgradeSummary :: Lens' NotificationSummary (Maybe LensUpgradeSummary)

Summary of lens upgrade.

notificationSummary_type :: Lens' NotificationSummary (Maybe NotificationType)

The type of notification.

PillarDifference

pillarDifference_pillarId :: Lens' PillarDifference (Maybe Text)

Undocumented member.

pillarDifference_questionDifferences :: Lens' PillarDifference (Maybe [QuestionDifference])

List of question differences.

pillarDifference_differenceStatus :: Lens' PillarDifference (Maybe DifferenceStatus)

Indicates the type of change to the pillar.

PillarReviewSummary

pillarReviewSummary_pillarId :: Lens' PillarReviewSummary (Maybe Text)

Undocumented member.

pillarReviewSummary_pillarName :: Lens' PillarReviewSummary (Maybe Text)

Undocumented member.

pillarReviewSummary_riskCounts :: Lens' PillarReviewSummary (Maybe (HashMap Risk Natural))

Undocumented member.

pillarReviewSummary_notes :: Lens' PillarReviewSummary (Maybe Text)

Undocumented member.

QuestionDifference

questionDifference_questionTitle :: Lens' QuestionDifference (Maybe Text)

Undocumented member.

questionDifference_differenceStatus :: Lens' QuestionDifference (Maybe DifferenceStatus)

Indicates the type of change to the question.

questionDifference_questionId :: Lens' QuestionDifference (Maybe Text)

Undocumented member.

ShareInvitation

shareInvitation_workloadId :: Lens' ShareInvitation (Maybe Text)

Undocumented member.

shareInvitation_shareInvitationId :: Lens' ShareInvitation (Maybe Text)

The ID assigned to the share invitation.

ShareInvitationSummary

shareInvitationSummary_sharedBy :: Lens' ShareInvitationSummary (Maybe Text)

Undocumented member.

shareInvitationSummary_sharedWith :: Lens' ShareInvitationSummary (Maybe Text)

Undocumented member.

shareInvitationSummary_permissionType :: Lens' ShareInvitationSummary (Maybe PermissionType)

Undocumented member.

shareInvitationSummary_workloadId :: Lens' ShareInvitationSummary (Maybe Text)

Undocumented member.

shareInvitationSummary_workloadName :: Lens' ShareInvitationSummary (Maybe Text)

Undocumented member.

shareInvitationSummary_shareInvitationId :: Lens' ShareInvitationSummary (Maybe Text)

The ID assigned to the share invitation.

VersionDifferences

versionDifferences_pillarDifferences :: Lens' VersionDifferences (Maybe [PillarDifference])

The differences between the base and latest versions of the lens.

Workload

workload_isReviewOwnerUpdateAcknowledged :: Lens' Workload (Maybe Bool)

Flag indicating whether the workload owner has acknowledged that the Review owner field is required.

If a Review owner is not added to the workload within 60 days of acknowledgement, access to the workload is restricted until an owner is added.

workload_architecturalDesign :: Lens' Workload (Maybe Text)

Undocumented member.

workload_accountIds :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_lenses :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_reviewRestrictionDate :: Lens' Workload (Maybe UTCTime)

Undocumented member.

workload_industry :: Lens' Workload (Maybe Text)

Undocumented member.

workload_environment :: Lens' Workload (Maybe WorkloadEnvironment)

Undocumented member.

workload_riskCounts :: Lens' Workload (Maybe (HashMap Risk Natural))

Undocumented member.

workload_awsRegions :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_owner :: Lens' Workload (Maybe Text)

Undocumented member.

workload_improvementStatus :: Lens' Workload (Maybe WorkloadImprovementStatus)

Undocumented member.

workload_workloadArn :: Lens' Workload (Maybe Text)

Undocumented member.

workload_industryType :: Lens' Workload (Maybe Text)

Undocumented member.

workload_workloadId :: Lens' Workload (Maybe Text)

Undocumented member.

workload_workloadName :: Lens' Workload (Maybe Text)

Undocumented member.

workload_updatedAt :: Lens' Workload (Maybe UTCTime)

Undocumented member.

workload_notes :: Lens' Workload (Maybe Text)

Undocumented member.

workload_reviewOwner :: Lens' Workload (Maybe Text)

Undocumented member.

workload_description :: Lens' Workload (Maybe Text)

Undocumented member.

workload_pillarPriorities :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_shareInvitationId :: Lens' Workload (Maybe Text)

The ID assigned to the share invitation.

workload_nonAwsRegions :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_tags :: Lens' Workload (Maybe (HashMap Text Text))

The tags associated with the workload.

WorkloadShare

workloadShare_status :: Lens' WorkloadShare (Maybe ShareStatus)

Undocumented member.

workloadShare_sharedBy :: Lens' WorkloadShare (Maybe Text)

Undocumented member.

workloadShare_sharedWith :: Lens' WorkloadShare (Maybe Text)

Undocumented member.

workloadShare_permissionType :: Lens' WorkloadShare (Maybe PermissionType)

Undocumented member.

workloadShare_workloadId :: Lens' WorkloadShare (Maybe Text)

Undocumented member.

workloadShare_workloadName :: Lens' WorkloadShare (Maybe Text)

Undocumented member.

workloadShare_shareId :: Lens' WorkloadShare (Maybe Text)

Undocumented member.

WorkloadShareSummary

workloadShareSummary_status :: Lens' WorkloadShareSummary (Maybe ShareStatus)

Undocumented member.

workloadShareSummary_sharedWith :: Lens' WorkloadShareSummary (Maybe Text)

Undocumented member.

workloadShareSummary_permissionType :: Lens' WorkloadShareSummary (Maybe PermissionType)

Undocumented member.

workloadShareSummary_shareId :: Lens' WorkloadShareSummary (Maybe Text)

Undocumented member.

WorkloadSummary

workloadSummary_lenses :: Lens' WorkloadSummary (Maybe [Text])

Undocumented member.

workloadSummary_riskCounts :: Lens' WorkloadSummary (Maybe (HashMap Risk Natural))

Undocumented member.

workloadSummary_owner :: Lens' WorkloadSummary (Maybe Text)

Undocumented member.

workloadSummary_improvementStatus :: Lens' WorkloadSummary (Maybe WorkloadImprovementStatus)

Undocumented member.

workloadSummary_workloadArn :: Lens' WorkloadSummary (Maybe Text)

Undocumented member.

workloadSummary_workloadId :: Lens' WorkloadSummary (Maybe Text)

Undocumented member.

workloadSummary_workloadName :: Lens' WorkloadSummary (Maybe Text)

Undocumented member.

workloadSummary_updatedAt :: Lens' WorkloadSummary (Maybe UTCTime)

Undocumented member.