Amazonka.WellArchitected.ListLensReviewImprovements

Description

List lens review improvements.

Creating a Request

data ListLensReviewImprovements

Input to list lens review improvements.

See: newListLensReviewImprovements smart constructor.

Constructors

ListLensReviewImprovements'

Fields

pillarId :: Maybe Text

milestoneNumber :: Maybe Natural

nextToken :: Maybe Text

maxResults :: Maybe Natural

The maximum number of results to return for this request.

workloadId :: Text

lensAlias :: Text

newListLensReviewImprovements

Arguments

:: Text
$sel:workloadId:ListLensReviewImprovements'

-> Text
$sel:lensAlias:ListLensReviewImprovements'

-> ListLensReviewImprovements

Create a value of ListLensReviewImprovements with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pillarId:ListLensReviewImprovements', listLensReviewImprovements_pillarId - Undocumented member.

$sel:milestoneNumber:ListLensReviewImprovements', listLensReviewImprovements_milestoneNumber - Undocumented member.

$sel:nextToken:ListLensReviewImprovements', listLensReviewImprovements_nextToken - Undocumented member.

$sel:maxResults:ListLensReviewImprovements', listLensReviewImprovements_maxResults - The maximum number of results to return for this request.

$sel:workloadId:ListLensReviewImprovements', listLensReviewImprovements_workloadId - Undocumented member.

$sel:lensAlias:ListLensReviewImprovements', listLensReviewImprovements_lensAlias - Undocumented member.

Request Lenses

listLensReviewImprovements_pillarId :: Lens' ListLensReviewImprovements (Maybe Text)

Undocumented member.

listLensReviewImprovements_milestoneNumber :: Lens' ListLensReviewImprovements (Maybe Natural)

Undocumented member.

listLensReviewImprovements_nextToken :: Lens' ListLensReviewImprovements (Maybe Text)

Undocumented member.

listLensReviewImprovements_maxResults :: Lens' ListLensReviewImprovements (Maybe Natural)

The maximum number of results to return for this request.

listLensReviewImprovements_workloadId :: Lens' ListLensReviewImprovements Text

Undocumented member.

listLensReviewImprovements_lensAlias :: Lens' ListLensReviewImprovements Text

Undocumented member.

Destructuring the Response

data ListLensReviewImprovementsResponse

Output of a list lens review improvements call.

See: newListLensReviewImprovementsResponse smart constructor.

Constructors

ListLensReviewImprovementsResponse'

Fields

improvementSummaries :: Maybe [ImprovementSummary]

lensAlias :: Maybe Text

milestoneNumber :: Maybe Natural

nextToken :: Maybe Text

workloadId :: Maybe Text

httpStatus :: Int

The response's http status code.

newListLensReviewImprovementsResponse

Arguments

:: Int
$sel:httpStatus:ListLensReviewImprovementsResponse'

-> ListLensReviewImprovementsResponse

Create a value of ListLensReviewImprovementsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:improvementSummaries:ListLensReviewImprovementsResponse', listLensReviewImprovementsResponse_improvementSummaries - Undocumented member.

$sel:lensAlias:ListLensReviewImprovements', listLensReviewImprovementsResponse_lensAlias - Undocumented member.

$sel:milestoneNumber:ListLensReviewImprovements', listLensReviewImprovementsResponse_milestoneNumber - Undocumented member.

$sel:nextToken:ListLensReviewImprovements', listLensReviewImprovementsResponse_nextToken - Undocumented member.

$sel:workloadId:ListLensReviewImprovements', listLensReviewImprovementsResponse_workloadId - Undocumented member.

$sel:httpStatus:ListLensReviewImprovementsResponse', listLensReviewImprovementsResponse_httpStatus - The response's http status code.

Response Lenses

listLensReviewImprovementsResponse_improvementSummaries :: Lens' ListLensReviewImprovementsResponse (Maybe [ImprovementSummary])

Undocumented member.

listLensReviewImprovementsResponse_lensAlias :: Lens' ListLensReviewImprovementsResponse (Maybe Text)

Undocumented member.

listLensReviewImprovementsResponse_milestoneNumber :: Lens' ListLensReviewImprovementsResponse (Maybe Natural)

Undocumented member.

listLensReviewImprovementsResponse_nextToken :: Lens' ListLensReviewImprovementsResponse (Maybe Text)

Undocumented member.

listLensReviewImprovementsResponse_workloadId :: Lens' ListLensReviewImprovementsResponse (Maybe Text)

Undocumented member.

listLensReviewImprovementsResponse_httpStatus :: Lens' ListLensReviewImprovementsResponse Int

The response's http status code.