Amazonka.WellArchitected.Types

Description

Service Configuration

defaultService :: Service

API version 2020-03-31 of the Amazon Well-Architected Tool SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The user input is not valid.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

User does not have sufficient access to perform this action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The resource already exists.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The user has reached their resource quota.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Request was denied due to request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

There is a problem with the AWS Well-Architected Tool API service.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource was not found.

AnswerReason

newtype AnswerReason

Constructors

AnswerReason'

Fields

fromAnswerReason :: Text

Bundled Patterns

pattern AnswerReason_ARCHITECTURE_CONSTRAINTS :: AnswerReason

pattern AnswerReason_BUSINESS_PRIORITIES :: AnswerReason

pattern AnswerReason_NONE :: AnswerReason

pattern AnswerReason_OTHER :: AnswerReason

pattern AnswerReason_OUT_OF_SCOPE :: AnswerReason

ChoiceReason

newtype ChoiceReason

Constructors

ChoiceReason'

Fields

fromChoiceReason :: Text

Bundled Patterns

pattern ChoiceReason_ARCHITECTURE_CONSTRAINTS :: ChoiceReason

pattern ChoiceReason_BUSINESS_PRIORITIES :: ChoiceReason

pattern ChoiceReason_NONE :: ChoiceReason

pattern ChoiceReason_OTHER :: ChoiceReason

pattern ChoiceReason_OUT_OF_SCOPE :: ChoiceReason

ChoiceStatus

newtype ChoiceStatus

Constructors

ChoiceStatus'

Fields

fromChoiceStatus :: Text

Bundled Patterns

pattern ChoiceStatus_NOT_APPLICABLE :: ChoiceStatus

pattern ChoiceStatus_SELECTED :: ChoiceStatus

pattern ChoiceStatus_UNSELECTED :: ChoiceStatus

DifferenceStatus

newtype DifferenceStatus

Constructors

DifferenceStatus'

Fields

fromDifferenceStatus :: Text

Bundled Patterns

pattern DifferenceStatus_DELETED :: DifferenceStatus

pattern DifferenceStatus_NEW :: DifferenceStatus

pattern DifferenceStatus_UPDATED :: DifferenceStatus

LensStatus

newtype LensStatus

Constructors

LensStatus'

Fields

fromLensStatus :: Text

Bundled Patterns

pattern LensStatus_CURRENT :: LensStatus

pattern LensStatus_DEPRECATED :: LensStatus

pattern LensStatus_NOT_CURRENT :: LensStatus

NotificationType

newtype NotificationType

Constructors

NotificationType'

Fields

fromNotificationType :: Text

Bundled Patterns

pattern NotificationType_LENS_VERSION_DEPRECATED :: NotificationType

pattern NotificationType_LENS_VERSION_UPGRADED :: NotificationType

PermissionType

newtype PermissionType

Permission granted on a workload share.

Constructors

PermissionType'

Fields

fromPermissionType :: Text

Bundled Patterns

pattern PermissionType_CONTRIBUTOR :: PermissionType

pattern PermissionType_READONLY :: PermissionType

Risk

newtype Risk

The risk for a given workload, lens review, pillar, or question.

Constructors

Risk'

Fields

fromRisk :: Text

Bundled Patterns

pattern Risk_HIGH :: Risk

pattern Risk_MEDIUM :: Risk

pattern Risk_NONE :: Risk

pattern Risk_NOT_APPLICABLE :: Risk

pattern Risk_UNANSWERED :: Risk

ShareInvitationAction

newtype ShareInvitationAction

Share invitation action taken by contributor.

Constructors

ShareInvitationAction'

Fields

fromShareInvitationAction :: Text

Bundled Patterns

pattern ShareInvitationAction_ACCEPT :: ShareInvitationAction

pattern ShareInvitationAction_REJECT :: ShareInvitationAction

ShareStatus

newtype ShareStatus

The status of a workload share.

Constructors

ShareStatus'

Fields

fromShareStatus :: Text

Bundled Patterns

pattern ShareStatus_ACCEPTED :: ShareStatus

pattern ShareStatus_EXPIRED :: ShareStatus

pattern ShareStatus_PENDING :: ShareStatus

pattern ShareStatus_REJECTED :: ShareStatus

pattern ShareStatus_REVOKED :: ShareStatus

WorkloadEnvironment

newtype WorkloadEnvironment

The environment for the workload.

Constructors

WorkloadEnvironment'

Fields

fromWorkloadEnvironment :: Text

Bundled Patterns

pattern WorkloadEnvironment_PREPRODUCTION :: WorkloadEnvironment

pattern WorkloadEnvironment_PRODUCTION :: WorkloadEnvironment

WorkloadImprovementStatus

newtype WorkloadImprovementStatus

The improvement status for a workload.

Constructors

WorkloadImprovementStatus'

Fields

fromWorkloadImprovementStatus :: Text

Bundled Patterns

pattern WorkloadImprovementStatus_COMPLETE :: WorkloadImprovementStatus

pattern WorkloadImprovementStatus_IN_PROGRESS :: WorkloadImprovementStatus

pattern WorkloadImprovementStatus_NOT_APPLICABLE :: WorkloadImprovementStatus

pattern WorkloadImprovementStatus_NOT_STARTED :: WorkloadImprovementStatus

pattern WorkloadImprovementStatus_RISK_ACKNOWLEDGED :: WorkloadImprovementStatus

Answer

data Answer

An answer of the question.

See: newAnswer smart constructor.

Constructors

Answer'

Fields

choiceAnswers :: Maybe [ChoiceAnswer]

A list of selected choices to a question in your workload.

helpfulResourceUrl :: Maybe Text

isApplicable :: Maybe Bool

pillarId :: Maybe Text

improvementPlanUrl :: Maybe Text

questionDescription :: Maybe Text

risk :: Maybe Risk

questionTitle :: Maybe Text

selectedChoices :: Maybe [Text]

reason :: Maybe AnswerReason

The reason why the question is not applicable to your workload.

choices :: Maybe [Choice]

questionId :: Maybe Text

notes :: Maybe Text

newAnswer :: Answer

Create a value of Answer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:choiceAnswers:Answer', answer_choiceAnswers - A list of selected choices to a question in your workload.

$sel:helpfulResourceUrl:Answer', answer_helpfulResourceUrl - Undocumented member.

$sel:isApplicable:Answer', answer_isApplicable - Undocumented member.

$sel:pillarId:Answer', answer_pillarId - Undocumented member.

$sel:improvementPlanUrl:Answer', answer_improvementPlanUrl - Undocumented member.

$sel:questionDescription:Answer', answer_questionDescription - Undocumented member.

$sel:risk:Answer', answer_risk - Undocumented member.

$sel:questionTitle:Answer', answer_questionTitle - Undocumented member.

$sel:selectedChoices:Answer', answer_selectedChoices - Undocumented member.

$sel:reason:Answer', answer_reason - The reason why the question is not applicable to your workload.

$sel:choices:Answer', answer_choices - Undocumented member.

$sel:questionId:Answer', answer_questionId - Undocumented member.

$sel:notes:Answer', answer_notes - Undocumented member.

answer_choiceAnswers :: Lens' Answer (Maybe [ChoiceAnswer])

A list of selected choices to a question in your workload.

answer_helpfulResourceUrl :: Lens' Answer (Maybe Text)

Undocumented member.

answer_isApplicable :: Lens' Answer (Maybe Bool)

Undocumented member.

answer_pillarId :: Lens' Answer (Maybe Text)

Undocumented member.

answer_improvementPlanUrl :: Lens' Answer (Maybe Text)

Undocumented member.

answer_questionDescription :: Lens' Answer (Maybe Text)

Undocumented member.

answer_risk :: Lens' Answer (Maybe Risk)

Undocumented member.

answer_questionTitle :: Lens' Answer (Maybe Text)

Undocumented member.

answer_selectedChoices :: Lens' Answer (Maybe [Text])

Undocumented member.

answer_reason :: Lens' Answer (Maybe AnswerReason)

The reason why the question is not applicable to your workload.

answer_choices :: Lens' Answer (Maybe [Choice])

Undocumented member.

answer_questionId :: Lens' Answer (Maybe Text)

Undocumented member.

answer_notes :: Lens' Answer (Maybe Text)

Undocumented member.

AnswerSummary

data AnswerSummary

An answer summary of a lens review in a workload.

See: newAnswerSummary smart constructor.

Constructors

AnswerSummary'

Fields

isApplicable :: Maybe Bool

pillarId :: Maybe Text

choiceAnswerSummaries :: Maybe [ChoiceAnswerSummary]

A list of selected choices to a question in your workload.

risk :: Maybe Risk

questionTitle :: Maybe Text

selectedChoices :: Maybe [Text]

reason :: Maybe AnswerReason

The reason why a choice is non-applicable to a question in your workload.

choices :: Maybe [Choice]

questionId :: Maybe Text

newAnswerSummary :: AnswerSummary

Create a value of AnswerSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isApplicable:AnswerSummary', answerSummary_isApplicable - Undocumented member.

$sel:pillarId:AnswerSummary', answerSummary_pillarId - Undocumented member.

$sel:choiceAnswerSummaries:AnswerSummary', answerSummary_choiceAnswerSummaries - A list of selected choices to a question in your workload.

$sel:risk:AnswerSummary', answerSummary_risk - Undocumented member.

$sel:questionTitle:AnswerSummary', answerSummary_questionTitle - Undocumented member.

$sel:selectedChoices:AnswerSummary', answerSummary_selectedChoices - Undocumented member.

$sel:reason:AnswerSummary', answerSummary_reason - The reason why a choice is non-applicable to a question in your workload.

$sel:choices:AnswerSummary', answerSummary_choices - Undocumented member.

$sel:questionId:AnswerSummary', answerSummary_questionId - Undocumented member.

answerSummary_isApplicable :: Lens' AnswerSummary (Maybe Bool)

Undocumented member.

answerSummary_pillarId :: Lens' AnswerSummary (Maybe Text)

Undocumented member.

answerSummary_choiceAnswerSummaries :: Lens' AnswerSummary (Maybe [ChoiceAnswerSummary])

A list of selected choices to a question in your workload.

answerSummary_risk :: Lens' AnswerSummary (Maybe Risk)

Undocumented member.

answerSummary_questionTitle :: Lens' AnswerSummary (Maybe Text)

Undocumented member.

answerSummary_selectedChoices :: Lens' AnswerSummary (Maybe [Text])

Undocumented member.

answerSummary_reason :: Lens' AnswerSummary (Maybe AnswerReason)

The reason why a choice is non-applicable to a question in your workload.

answerSummary_choices :: Lens' AnswerSummary (Maybe [Choice])

Undocumented member.

answerSummary_questionId :: Lens' AnswerSummary (Maybe Text)

Undocumented member.

Choice

data Choice

A choice available to answer question.

See: newChoice smart constructor.

Constructors

Choice'

Fields

title :: Maybe Text

description :: Maybe Text

choiceId :: Maybe Text

newChoice :: Choice

Create a value of Choice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:title:Choice', choice_title - Undocumented member.

$sel:description:Choice', choice_description - Undocumented member.

$sel:choiceId:Choice', choice_choiceId - Undocumented member.

choice_title :: Lens' Choice (Maybe Text)

Undocumented member.

choice_description :: Lens' Choice (Maybe Text)

Undocumented member.

choice_choiceId :: Lens' Choice (Maybe Text)

Undocumented member.

ChoiceAnswer

data ChoiceAnswer

A choice that has been answered on a question in your workload.

See: newChoiceAnswer smart constructor.

Constructors

ChoiceAnswer'

Fields

status :: Maybe ChoiceStatus

The status of a choice.

reason :: Maybe ChoiceReason

The reason why a choice is non-applicable to a question in your workload.

notes :: Maybe Text

The notes associated with a choice.

choiceId :: Maybe Text

newChoiceAnswer :: ChoiceAnswer

Create a value of ChoiceAnswer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ChoiceAnswer', choiceAnswer_status - The status of a choice.

$sel:reason:ChoiceAnswer', choiceAnswer_reason - The reason why a choice is non-applicable to a question in your workload.

$sel:notes:ChoiceAnswer', choiceAnswer_notes - The notes associated with a choice.

$sel:choiceId:ChoiceAnswer', choiceAnswer_choiceId - Undocumented member.

choiceAnswer_status :: Lens' ChoiceAnswer (Maybe ChoiceStatus)

The status of a choice.

choiceAnswer_reason :: Lens' ChoiceAnswer (Maybe ChoiceReason)

The reason why a choice is non-applicable to a question in your workload.

choiceAnswer_notes :: Lens' ChoiceAnswer (Maybe Text)

The notes associated with a choice.

choiceAnswer_choiceId :: Lens' ChoiceAnswer (Maybe Text)

Undocumented member.

ChoiceAnswerSummary

data ChoiceAnswerSummary

A choice summary that has been answered on a question in your workload.

See: newChoiceAnswerSummary smart constructor.

Constructors

ChoiceAnswerSummary'

Fields

status :: Maybe ChoiceStatus

The status of a choice.

reason :: Maybe ChoiceReason

The reason why a choice is non-applicable to a question in your workload.

choiceId :: Maybe Text

newChoiceAnswerSummary :: ChoiceAnswerSummary

Create a value of ChoiceAnswerSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ChoiceAnswerSummary', choiceAnswerSummary_status - The status of a choice.

$sel:reason:ChoiceAnswerSummary', choiceAnswerSummary_reason - The reason why a choice is non-applicable to a question in your workload.

$sel:choiceId:ChoiceAnswerSummary', choiceAnswerSummary_choiceId - Undocumented member.

choiceAnswerSummary_status :: Lens' ChoiceAnswerSummary (Maybe ChoiceStatus)

The status of a choice.

choiceAnswerSummary_reason :: Lens' ChoiceAnswerSummary (Maybe ChoiceReason)

The reason why a choice is non-applicable to a question in your workload.

choiceAnswerSummary_choiceId :: Lens' ChoiceAnswerSummary (Maybe Text)

Undocumented member.

ChoiceUpdate

data ChoiceUpdate

A list of choices to be updated.

See: newChoiceUpdate smart constructor.

Constructors

ChoiceUpdate'

Fields

reason :: Maybe ChoiceReason

The reason why a choice is non-applicable to a question in your workload.

notes :: Maybe Text

The notes associated with a choice.

status :: ChoiceStatus

The status of a choice.

newChoiceUpdate

Arguments

:: ChoiceStatus
$sel:status:ChoiceUpdate'

-> ChoiceUpdate

Create a value of ChoiceUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reason:ChoiceUpdate', choiceUpdate_reason - The reason why a choice is non-applicable to a question in your workload.

$sel:notes:ChoiceUpdate', choiceUpdate_notes - The notes associated with a choice.

$sel:status:ChoiceUpdate', choiceUpdate_status - The status of a choice.

choiceUpdate_reason :: Lens' ChoiceUpdate (Maybe ChoiceReason)

The reason why a choice is non-applicable to a question in your workload.

choiceUpdate_notes :: Lens' ChoiceUpdate (Maybe Text)

The notes associated with a choice.

choiceUpdate_status :: Lens' ChoiceUpdate ChoiceStatus

The status of a choice.

ImprovementSummary

data ImprovementSummary

An improvement summary of a lens review in a workload.

See: newImprovementSummary smart constructor.

Constructors

ImprovementSummary'

Fields

pillarId :: Maybe Text

improvementPlanUrl :: Maybe Text

risk :: Maybe Risk

questionTitle :: Maybe Text

questionId :: Maybe Text

newImprovementSummary :: ImprovementSummary

Create a value of ImprovementSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pillarId:ImprovementSummary', improvementSummary_pillarId - Undocumented member.

$sel:improvementPlanUrl:ImprovementSummary', improvementSummary_improvementPlanUrl - Undocumented member.

$sel:risk:ImprovementSummary', improvementSummary_risk - Undocumented member.

$sel:questionTitle:ImprovementSummary', improvementSummary_questionTitle - Undocumented member.

$sel:questionId:ImprovementSummary', improvementSummary_questionId - Undocumented member.

improvementSummary_pillarId :: Lens' ImprovementSummary (Maybe Text)

Undocumented member.

improvementSummary_improvementPlanUrl :: Lens' ImprovementSummary (Maybe Text)

Undocumented member.

improvementSummary_risk :: Lens' ImprovementSummary (Maybe Risk)

Undocumented member.

improvementSummary_questionTitle :: Lens' ImprovementSummary (Maybe Text)

Undocumented member.

improvementSummary_questionId :: Lens' ImprovementSummary (Maybe Text)

Undocumented member.

LensReview

data LensReview

A lens review of a question.

See: newLensReview smart constructor.

Constructors

LensReview'

Fields

lensAlias :: Maybe Text

riskCounts :: Maybe (HashMap Risk Natural)

lensName :: Maybe Text

nextToken :: Maybe Text

pillarReviewSummaries :: Maybe [PillarReviewSummary]

updatedAt :: Maybe POSIX

lensStatus :: Maybe LensStatus

The status of the lens.

notes :: Maybe Text

lensVersion :: Maybe Text

The version of the lens.

newLensReview :: LensReview

Create a value of LensReview with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lensAlias:LensReview', lensReview_lensAlias - Undocumented member.

$sel:riskCounts:LensReview', lensReview_riskCounts - Undocumented member.

$sel:lensName:LensReview', lensReview_lensName - Undocumented member.

$sel:nextToken:LensReview', lensReview_nextToken - Undocumented member.

$sel:pillarReviewSummaries:LensReview', lensReview_pillarReviewSummaries - Undocumented member.

$sel:updatedAt:LensReview', lensReview_updatedAt - Undocumented member.

$sel:lensStatus:LensReview', lensReview_lensStatus - The status of the lens.

$sel:notes:LensReview', lensReview_notes - Undocumented member.

$sel:lensVersion:LensReview', lensReview_lensVersion - The version of the lens.

lensReview_lensAlias :: Lens' LensReview (Maybe Text)

Undocumented member.

lensReview_riskCounts :: Lens' LensReview (Maybe (HashMap Risk Natural))

Undocumented member.

lensReview_lensName :: Lens' LensReview (Maybe Text)

Undocumented member.

lensReview_nextToken :: Lens' LensReview (Maybe Text)

Undocumented member.

lensReview_pillarReviewSummaries :: Lens' LensReview (Maybe [PillarReviewSummary])

Undocumented member.

lensReview_updatedAt :: Lens' LensReview (Maybe UTCTime)

Undocumented member.

lensReview_lensStatus :: Lens' LensReview (Maybe LensStatus)

The status of the lens.

lensReview_notes :: Lens' LensReview (Maybe Text)

Undocumented member.

lensReview_lensVersion :: Lens' LensReview (Maybe Text)

The version of the lens.

LensReviewReport

data LensReviewReport

A report of a lens review.

See: newLensReviewReport smart constructor.

Constructors

LensReviewReport'

Fields

lensAlias :: Maybe Text

base64String :: Maybe Text

newLensReviewReport :: LensReviewReport

Create a value of LensReviewReport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lensAlias:LensReviewReport', lensReviewReport_lensAlias - Undocumented member.

$sel:base64String:LensReviewReport', lensReviewReport_base64String - Undocumented member.

lensReviewReport_lensAlias :: Lens' LensReviewReport (Maybe Text)

Undocumented member.

lensReviewReport_base64String :: Lens' LensReviewReport (Maybe Text)

Undocumented member.

LensReviewSummary

data LensReviewSummary

A lens review summary of a workload.

See: newLensReviewSummary smart constructor.

Constructors

LensReviewSummary'

Fields

lensAlias :: Maybe Text

riskCounts :: Maybe (HashMap Risk Natural)

lensName :: Maybe Text

updatedAt :: Maybe POSIX

lensStatus :: Maybe LensStatus

The status of the lens.

lensVersion :: Maybe Text

The version of the lens.

newLensReviewSummary :: LensReviewSummary

Create a value of LensReviewSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lensAlias:LensReviewSummary', lensReviewSummary_lensAlias - Undocumented member.

$sel:riskCounts:LensReviewSummary', lensReviewSummary_riskCounts - Undocumented member.

$sel:lensName:LensReviewSummary', lensReviewSummary_lensName - Undocumented member.

$sel:updatedAt:LensReviewSummary', lensReviewSummary_updatedAt - Undocumented member.

$sel:lensStatus:LensReviewSummary', lensReviewSummary_lensStatus - The status of the lens.

$sel:lensVersion:LensReviewSummary', lensReviewSummary_lensVersion - The version of the lens.

lensReviewSummary_lensAlias :: Lens' LensReviewSummary (Maybe Text)

Undocumented member.

lensReviewSummary_riskCounts :: Lens' LensReviewSummary (Maybe (HashMap Risk Natural))

Undocumented member.

lensReviewSummary_lensName :: Lens' LensReviewSummary (Maybe Text)

Undocumented member.

lensReviewSummary_updatedAt :: Lens' LensReviewSummary (Maybe UTCTime)

Undocumented member.

lensReviewSummary_lensStatus :: Lens' LensReviewSummary (Maybe LensStatus)

The status of the lens.

lensReviewSummary_lensVersion :: Lens' LensReviewSummary (Maybe Text)

The version of the lens.

LensSummary

data LensSummary

A lens summary of a lens.

See: newLensSummary smart constructor.

Constructors

LensSummary'

Fields

lensAlias :: Maybe Text

lensName :: Maybe Text

lensVersion :: Maybe Text

The version of the lens.

description :: Maybe Text

newLensSummary :: LensSummary

Create a value of LensSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lensAlias:LensSummary', lensSummary_lensAlias - Undocumented member.

$sel:lensName:LensSummary', lensSummary_lensName - Undocumented member.

$sel:lensVersion:LensSummary', lensSummary_lensVersion - The version of the lens.

$sel:description:LensSummary', lensSummary_description - Undocumented member.

lensSummary_lensAlias :: Lens' LensSummary (Maybe Text)

Undocumented member.

lensSummary_lensName :: Lens' LensSummary (Maybe Text)

Undocumented member.

lensSummary_lensVersion :: Lens' LensSummary (Maybe Text)

The version of the lens.

lensSummary_description :: Lens' LensSummary (Maybe Text)

Undocumented member.

LensUpgradeSummary

data LensUpgradeSummary

Lens upgrade summary return object.

See: newLensUpgradeSummary smart constructor.

Constructors

LensUpgradeSummary'

Fields

lensAlias :: Maybe Text

latestLensVersion :: Maybe Text

The latest version of the lens.

currentLensVersion :: Maybe Text

The current version of the lens.

workloadId :: Maybe Text

workloadName :: Maybe Text

newLensUpgradeSummary :: LensUpgradeSummary

Create a value of LensUpgradeSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lensAlias:LensUpgradeSummary', lensUpgradeSummary_lensAlias - Undocumented member.

$sel:latestLensVersion:LensUpgradeSummary', lensUpgradeSummary_latestLensVersion - The latest version of the lens.

$sel:currentLensVersion:LensUpgradeSummary', lensUpgradeSummary_currentLensVersion - The current version of the lens.

$sel:workloadId:LensUpgradeSummary', lensUpgradeSummary_workloadId - Undocumented member.

$sel:workloadName:LensUpgradeSummary', lensUpgradeSummary_workloadName - Undocumented member.

lensUpgradeSummary_lensAlias :: Lens' LensUpgradeSummary (Maybe Text)

Undocumented member.

lensUpgradeSummary_latestLensVersion :: Lens' LensUpgradeSummary (Maybe Text)

The latest version of the lens.

lensUpgradeSummary_currentLensVersion :: Lens' LensUpgradeSummary (Maybe Text)

The current version of the lens.

lensUpgradeSummary_workloadId :: Lens' LensUpgradeSummary (Maybe Text)

Undocumented member.

lensUpgradeSummary_workloadName :: Lens' LensUpgradeSummary (Maybe Text)

Undocumented member.

Milestone

data Milestone

A milestone return object.

See: newMilestone smart constructor.

Constructors

Milestone'

Fields

workload :: Maybe Workload

milestoneNumber :: Maybe Natural

milestoneName :: Maybe Text

recordedAt :: Maybe POSIX

newMilestone :: Milestone

Create a value of Milestone with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workload:Milestone', milestone_workload - Undocumented member.

$sel:milestoneNumber:Milestone', milestone_milestoneNumber - Undocumented member.

$sel:milestoneName:Milestone', milestone_milestoneName - Undocumented member.

$sel:recordedAt:Milestone', milestone_recordedAt - Undocumented member.

milestone_workload :: Lens' Milestone (Maybe Workload)

Undocumented member.

milestone_milestoneNumber :: Lens' Milestone (Maybe Natural)

Undocumented member.

milestone_milestoneName :: Lens' Milestone (Maybe Text)

Undocumented member.

milestone_recordedAt :: Lens' Milestone (Maybe UTCTime)

Undocumented member.

MilestoneSummary

data MilestoneSummary

A milestone summary return object.

See: newMilestoneSummary smart constructor.

Constructors

MilestoneSummary'

Fields

milestoneNumber :: Maybe Natural

milestoneName :: Maybe Text

recordedAt :: Maybe POSIX

workloadSummary :: Maybe WorkloadSummary

newMilestoneSummary :: MilestoneSummary

Create a value of MilestoneSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:milestoneNumber:MilestoneSummary', milestoneSummary_milestoneNumber - Undocumented member.

$sel:milestoneName:MilestoneSummary', milestoneSummary_milestoneName - Undocumented member.

$sel:recordedAt:MilestoneSummary', milestoneSummary_recordedAt - Undocumented member.

$sel:workloadSummary:MilestoneSummary', milestoneSummary_workloadSummary - Undocumented member.

milestoneSummary_milestoneNumber :: Lens' MilestoneSummary (Maybe Natural)

Undocumented member.

milestoneSummary_milestoneName :: Lens' MilestoneSummary (Maybe Text)

Undocumented member.

milestoneSummary_recordedAt :: Lens' MilestoneSummary (Maybe UTCTime)

Undocumented member.

milestoneSummary_workloadSummary :: Lens' MilestoneSummary (Maybe WorkloadSummary)

Undocumented member.

NotificationSummary

data NotificationSummary

A notification summary return object.

See: newNotificationSummary smart constructor.

Constructors

NotificationSummary'

Fields

lensUpgradeSummary :: Maybe LensUpgradeSummary

Summary of lens upgrade.

type' :: Maybe NotificationType

The type of notification.

newNotificationSummary :: NotificationSummary

Create a value of NotificationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lensUpgradeSummary:NotificationSummary', notificationSummary_lensUpgradeSummary - Summary of lens upgrade.

$sel:type':NotificationSummary', notificationSummary_type - The type of notification.

notificationSummary_lensUpgradeSummary :: Lens' NotificationSummary (Maybe LensUpgradeSummary)

Summary of lens upgrade.

notificationSummary_type :: Lens' NotificationSummary (Maybe NotificationType)

The type of notification.

PillarDifference

data PillarDifference

A pillar difference return object.

See: newPillarDifference smart constructor.

Constructors

PillarDifference'

Fields

pillarId :: Maybe Text

questionDifferences :: Maybe [QuestionDifference]

List of question differences.

differenceStatus :: Maybe DifferenceStatus

Indicates the type of change to the pillar.

newPillarDifference :: PillarDifference

Create a value of PillarDifference with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pillarId:PillarDifference', pillarDifference_pillarId - Undocumented member.

$sel:questionDifferences:PillarDifference', pillarDifference_questionDifferences - List of question differences.

$sel:differenceStatus:PillarDifference', pillarDifference_differenceStatus - Indicates the type of change to the pillar.

pillarDifference_pillarId :: Lens' PillarDifference (Maybe Text)

Undocumented member.

pillarDifference_questionDifferences :: Lens' PillarDifference (Maybe [QuestionDifference])

List of question differences.

pillarDifference_differenceStatus :: Lens' PillarDifference (Maybe DifferenceStatus)

Indicates the type of change to the pillar.

PillarReviewSummary

data PillarReviewSummary

A pillar review summary of a lens review.

See: newPillarReviewSummary smart constructor.

Constructors

PillarReviewSummary'

Fields

pillarId :: Maybe Text

pillarName :: Maybe Text

riskCounts :: Maybe (HashMap Risk Natural)

notes :: Maybe Text

newPillarReviewSummary :: PillarReviewSummary

Create a value of PillarReviewSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pillarId:PillarReviewSummary', pillarReviewSummary_pillarId - Undocumented member.

$sel:pillarName:PillarReviewSummary', pillarReviewSummary_pillarName - Undocumented member.

$sel:riskCounts:PillarReviewSummary', pillarReviewSummary_riskCounts - Undocumented member.

$sel:notes:PillarReviewSummary', pillarReviewSummary_notes - Undocumented member.

pillarReviewSummary_pillarId :: Lens' PillarReviewSummary (Maybe Text)

Undocumented member.

pillarReviewSummary_pillarName :: Lens' PillarReviewSummary (Maybe Text)

Undocumented member.

pillarReviewSummary_riskCounts :: Lens' PillarReviewSummary (Maybe (HashMap Risk Natural))

Undocumented member.

pillarReviewSummary_notes :: Lens' PillarReviewSummary (Maybe Text)

Undocumented member.

QuestionDifference

data QuestionDifference

A question difference return object.

See: newQuestionDifference smart constructor.

Constructors

QuestionDifference'

Fields

questionTitle :: Maybe Text

differenceStatus :: Maybe DifferenceStatus

Indicates the type of change to the question.

questionId :: Maybe Text

newQuestionDifference :: QuestionDifference

Create a value of QuestionDifference with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:questionTitle:QuestionDifference', questionDifference_questionTitle - Undocumented member.

$sel:differenceStatus:QuestionDifference', questionDifference_differenceStatus - Indicates the type of change to the question.

$sel:questionId:QuestionDifference', questionDifference_questionId - Undocumented member.

questionDifference_questionTitle :: Lens' QuestionDifference (Maybe Text)

Undocumented member.

questionDifference_differenceStatus :: Lens' QuestionDifference (Maybe DifferenceStatus)

Indicates the type of change to the question.

questionDifference_questionId :: Lens' QuestionDifference (Maybe Text)

Undocumented member.

ShareInvitation

data ShareInvitation

The share invitation.

See: newShareInvitation smart constructor.

Constructors

ShareInvitation'

Fields

workloadId :: Maybe Text

shareInvitationId :: Maybe Text

The ID assigned to the share invitation.

newShareInvitation :: ShareInvitation

Create a value of ShareInvitation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workloadId:ShareInvitation', shareInvitation_workloadId - Undocumented member.

$sel:shareInvitationId:ShareInvitation', shareInvitation_shareInvitationId - The ID assigned to the share invitation.

shareInvitation_workloadId :: Lens' ShareInvitation (Maybe Text)

Undocumented member.

shareInvitation_shareInvitationId :: Lens' ShareInvitation (Maybe Text)

The ID assigned to the share invitation.

ShareInvitationSummary

data ShareInvitationSummary

A share invitation summary return object.

See: newShareInvitationSummary smart constructor.

Constructors

ShareInvitationSummary'

Fields

sharedBy :: Maybe Text

sharedWith :: Maybe Text

permissionType :: Maybe PermissionType

workloadId :: Maybe Text

workloadName :: Maybe Text

shareInvitationId :: Maybe Text

The ID assigned to the share invitation.

newShareInvitationSummary :: ShareInvitationSummary

Create a value of ShareInvitationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sharedBy:ShareInvitationSummary', shareInvitationSummary_sharedBy - Undocumented member.

$sel:sharedWith:ShareInvitationSummary', shareInvitationSummary_sharedWith - Undocumented member.

$sel:permissionType:ShareInvitationSummary', shareInvitationSummary_permissionType - Undocumented member.

$sel:workloadId:ShareInvitationSummary', shareInvitationSummary_workloadId - Undocumented member.

$sel:workloadName:ShareInvitationSummary', shareInvitationSummary_workloadName - Undocumented member.

$sel:shareInvitationId:ShareInvitationSummary', shareInvitationSummary_shareInvitationId - The ID assigned to the share invitation.

shareInvitationSummary_sharedBy :: Lens' ShareInvitationSummary (Maybe Text)

Undocumented member.

shareInvitationSummary_sharedWith :: Lens' ShareInvitationSummary (Maybe Text)

Undocumented member.

shareInvitationSummary_permissionType :: Lens' ShareInvitationSummary (Maybe PermissionType)

Undocumented member.

shareInvitationSummary_workloadId :: Lens' ShareInvitationSummary (Maybe Text)

Undocumented member.

shareInvitationSummary_workloadName :: Lens' ShareInvitationSummary (Maybe Text)

Undocumented member.

shareInvitationSummary_shareInvitationId :: Lens' ShareInvitationSummary (Maybe Text)

The ID assigned to the share invitation.

VersionDifferences

data VersionDifferences

The differences between the base and latest versions of the lens.

See: newVersionDifferences smart constructor.

Constructors

VersionDifferences'

Fields

pillarDifferences :: Maybe [PillarDifference]

The differences between the base and latest versions of the lens.

newVersionDifferences :: VersionDifferences

Create a value of VersionDifferences with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pillarDifferences:VersionDifferences', versionDifferences_pillarDifferences - The differences between the base and latest versions of the lens.

versionDifferences_pillarDifferences :: Lens' VersionDifferences (Maybe [PillarDifference])

The differences between the base and latest versions of the lens.

Workload

data Workload

A workload return object.

See: newWorkload smart constructor.

Constructors

Workload'

Fields

isReviewOwnerUpdateAcknowledged :: Maybe Bool

Flag indicating whether the workload owner has acknowledged that the Review owner field is required.

If a Review owner is not added to the workload within 60 days of acknowledgement, access to the workload is restricted until an owner is added.

architecturalDesign :: Maybe Text

accountIds :: Maybe [Text]

lenses :: Maybe [Text]

reviewRestrictionDate :: Maybe POSIX

industry :: Maybe Text

environment :: Maybe WorkloadEnvironment

riskCounts :: Maybe (HashMap Risk Natural)

awsRegions :: Maybe [Text]

owner :: Maybe Text

improvementStatus :: Maybe WorkloadImprovementStatus

workloadArn :: Maybe Text

industryType :: Maybe Text

workloadId :: Maybe Text

workloadName :: Maybe Text

updatedAt :: Maybe POSIX

notes :: Maybe Text

reviewOwner :: Maybe Text

description :: Maybe Text

pillarPriorities :: Maybe [Text]

shareInvitationId :: Maybe Text

The ID assigned to the share invitation.

nonAwsRegions :: Maybe [Text]

tags :: Maybe (HashMap Text Text)

The tags associated with the workload.

newWorkload :: Workload

Create a value of Workload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isReviewOwnerUpdateAcknowledged:Workload', workload_isReviewOwnerUpdateAcknowledged - Flag indicating whether the workload owner has acknowledged that the Review owner field is required.

If a Review owner is not added to the workload within 60 days of acknowledgement, access to the workload is restricted until an owner is added.

$sel:architecturalDesign:Workload', workload_architecturalDesign - Undocumented member.

$sel:accountIds:Workload', workload_accountIds - Undocumented member.

$sel:lenses:Workload', workload_lenses - Undocumented member.

$sel:reviewRestrictionDate:Workload', workload_reviewRestrictionDate - Undocumented member.

$sel:industry:Workload', workload_industry - Undocumented member.

$sel:environment:Workload', workload_environment - Undocumented member.

$sel:riskCounts:Workload', workload_riskCounts - Undocumented member.

$sel:awsRegions:Workload', workload_awsRegions - Undocumented member.

$sel:owner:Workload', workload_owner - Undocumented member.

$sel:improvementStatus:Workload', workload_improvementStatus - Undocumented member.

$sel:workloadArn:Workload', workload_workloadArn - Undocumented member.

$sel:industryType:Workload', workload_industryType - Undocumented member.

$sel:workloadId:Workload', workload_workloadId - Undocumented member.

$sel:workloadName:Workload', workload_workloadName - Undocumented member.

$sel:updatedAt:Workload', workload_updatedAt - Undocumented member.

$sel:notes:Workload', workload_notes - Undocumented member.

$sel:reviewOwner:Workload', workload_reviewOwner - Undocumented member.

$sel:description:Workload', workload_description - Undocumented member.

$sel:pillarPriorities:Workload', workload_pillarPriorities - Undocumented member.

$sel:shareInvitationId:Workload', workload_shareInvitationId - The ID assigned to the share invitation.

$sel:nonAwsRegions:Workload', workload_nonAwsRegions - Undocumented member.

$sel:tags:Workload', workload_tags - The tags associated with the workload.

workload_isReviewOwnerUpdateAcknowledged :: Lens' Workload (Maybe Bool)

Flag indicating whether the workload owner has acknowledged that the Review owner field is required.

If a Review owner is not added to the workload within 60 days of acknowledgement, access to the workload is restricted until an owner is added.

workload_architecturalDesign :: Lens' Workload (Maybe Text)

Undocumented member.

workload_accountIds :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_lenses :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_reviewRestrictionDate :: Lens' Workload (Maybe UTCTime)

Undocumented member.

workload_industry :: Lens' Workload (Maybe Text)

Undocumented member.

workload_environment :: Lens' Workload (Maybe WorkloadEnvironment)

Undocumented member.

workload_riskCounts :: Lens' Workload (Maybe (HashMap Risk Natural))

Undocumented member.

workload_awsRegions :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_owner :: Lens' Workload (Maybe Text)

Undocumented member.

workload_improvementStatus :: Lens' Workload (Maybe WorkloadImprovementStatus)

Undocumented member.

workload_workloadArn :: Lens' Workload (Maybe Text)

Undocumented member.

workload_industryType :: Lens' Workload (Maybe Text)

Undocumented member.

workload_workloadId :: Lens' Workload (Maybe Text)

Undocumented member.

workload_workloadName :: Lens' Workload (Maybe Text)

Undocumented member.

workload_updatedAt :: Lens' Workload (Maybe UTCTime)

Undocumented member.

workload_notes :: Lens' Workload (Maybe Text)

Undocumented member.

workload_reviewOwner :: Lens' Workload (Maybe Text)

Undocumented member.

workload_description :: Lens' Workload (Maybe Text)

Undocumented member.

workload_pillarPriorities :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_shareInvitationId :: Lens' Workload (Maybe Text)

The ID assigned to the share invitation.

workload_nonAwsRegions :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_tags :: Lens' Workload (Maybe (HashMap Text Text))

The tags associated with the workload.

WorkloadShare

data WorkloadShare

A workload share return object.

See: newWorkloadShare smart constructor.

Constructors

WorkloadShare'

Fields

status :: Maybe ShareStatus

sharedBy :: Maybe Text

sharedWith :: Maybe Text

permissionType :: Maybe PermissionType

workloadId :: Maybe Text

workloadName :: Maybe Text

shareId :: Maybe Text

newWorkloadShare :: WorkloadShare

Create a value of WorkloadShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:WorkloadShare', workloadShare_status - Undocumented member.

$sel:sharedBy:WorkloadShare', workloadShare_sharedBy - Undocumented member.

$sel:sharedWith:WorkloadShare', workloadShare_sharedWith - Undocumented member.

$sel:permissionType:WorkloadShare', workloadShare_permissionType - Undocumented member.

$sel:workloadId:WorkloadShare', workloadShare_workloadId - Undocumented member.

$sel:workloadName:WorkloadShare', workloadShare_workloadName - Undocumented member.

$sel:shareId:WorkloadShare', workloadShare_shareId - Undocumented member.

workloadShare_status :: Lens' WorkloadShare (Maybe ShareStatus)

Undocumented member.

workloadShare_sharedBy :: Lens' WorkloadShare (Maybe Text)

Undocumented member.

workloadShare_sharedWith :: Lens' WorkloadShare (Maybe Text)

Undocumented member.

workloadShare_permissionType :: Lens' WorkloadShare (Maybe PermissionType)

Undocumented member.

workloadShare_workloadId :: Lens' WorkloadShare (Maybe Text)

Undocumented member.

workloadShare_workloadName :: Lens' WorkloadShare (Maybe Text)

Undocumented member.

workloadShare_shareId :: Lens' WorkloadShare (Maybe Text)

Undocumented member.

WorkloadShareSummary

data WorkloadShareSummary

A workload share summary return object.

See: newWorkloadShareSummary smart constructor.

Constructors

WorkloadShareSummary'

Fields

status :: Maybe ShareStatus

sharedWith :: Maybe Text

permissionType :: Maybe PermissionType

shareId :: Maybe Text

newWorkloadShareSummary :: WorkloadShareSummary

Create a value of WorkloadShareSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:WorkloadShareSummary', workloadShareSummary_status - Undocumented member.

$sel:sharedWith:WorkloadShareSummary', workloadShareSummary_sharedWith - Undocumented member.

$sel:permissionType:WorkloadShareSummary', workloadShareSummary_permissionType - Undocumented member.

$sel:shareId:WorkloadShareSummary', workloadShareSummary_shareId - Undocumented member.

workloadShareSummary_status :: Lens' WorkloadShareSummary (Maybe ShareStatus)

Undocumented member.

workloadShareSummary_sharedWith :: Lens' WorkloadShareSummary (Maybe Text)

Undocumented member.

workloadShareSummary_permissionType :: Lens' WorkloadShareSummary (Maybe PermissionType)

Undocumented member.

workloadShareSummary_shareId :: Lens' WorkloadShareSummary (Maybe Text)

Undocumented member.

WorkloadSummary

data WorkloadSummary

A workload summary return object.

See: newWorkloadSummary smart constructor.

Constructors

WorkloadSummary'

Fields

lenses :: Maybe [Text]

riskCounts :: Maybe (HashMap Risk Natural)

owner :: Maybe Text

improvementStatus :: Maybe WorkloadImprovementStatus

workloadArn :: Maybe Text

workloadId :: Maybe Text

workloadName :: Maybe Text

updatedAt :: Maybe POSIX

newWorkloadSummary :: WorkloadSummary

Create a value of WorkloadSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lenses:WorkloadSummary', workloadSummary_lenses - Undocumented member.

$sel:riskCounts:WorkloadSummary', workloadSummary_riskCounts - Undocumented member.

$sel:owner:WorkloadSummary', workloadSummary_owner - Undocumented member.

$sel:improvementStatus:WorkloadSummary', workloadSummary_improvementStatus - Undocumented member.

$sel:workloadArn:WorkloadSummary', workloadSummary_workloadArn - Undocumented member.

$sel:workloadId:WorkloadSummary', workloadSummary_workloadId - Undocumented member.

$sel:workloadName:WorkloadSummary', workloadSummary_workloadName - Undocumented member.

$sel:updatedAt:WorkloadSummary', workloadSummary_updatedAt - Undocumented member.

workloadSummary_lenses :: Lens' WorkloadSummary (Maybe [Text])

Undocumented member.

workloadSummary_riskCounts :: Lens' WorkloadSummary (Maybe (HashMap Risk Natural))

Undocumented member.

workloadSummary_owner :: Lens' WorkloadSummary (Maybe Text)

Undocumented member.

workloadSummary_improvementStatus :: Lens' WorkloadSummary (Maybe WorkloadImprovementStatus)

Undocumented member.

workloadSummary_workloadArn :: Lens' WorkloadSummary (Maybe Text)

Undocumented member.

workloadSummary_workloadId :: Lens' WorkloadSummary (Maybe Text)

Undocumented member.

workloadSummary_workloadName :: Lens' WorkloadSummary (Maybe Text)

Undocumented member.

workloadSummary_updatedAt :: Lens' WorkloadSummary (Maybe UTCTime)

Undocumented member.