Amazonka.WellArchitected.Types.ImprovementSummary

Description

Documentation

data ImprovementSummary

An improvement summary of a lens review in a workload.

See: newImprovementSummary smart constructor.

Constructors

ImprovementSummary'

Fields

pillarId :: Maybe Text

improvementPlanUrl :: Maybe Text

risk :: Maybe Risk

questionTitle :: Maybe Text

questionId :: Maybe Text

newImprovementSummary :: ImprovementSummary

Create a value of ImprovementSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pillarId:ImprovementSummary', improvementSummary_pillarId - Undocumented member.

$sel:improvementPlanUrl:ImprovementSummary', improvementSummary_improvementPlanUrl - Undocumented member.

$sel:risk:ImprovementSummary', improvementSummary_risk - Undocumented member.

$sel:questionTitle:ImprovementSummary', improvementSummary_questionTitle - Undocumented member.

$sel:questionId:ImprovementSummary', improvementSummary_questionId - Undocumented member.

improvementSummary_pillarId :: Lens' ImprovementSummary (Maybe Text)

Undocumented member.

improvementSummary_improvementPlanUrl :: Lens' ImprovementSummary (Maybe Text)

Undocumented member.

improvementSummary_risk :: Lens' ImprovementSummary (Maybe Risk)

Undocumented member.

improvementSummary_questionTitle :: Lens' ImprovementSummary (Maybe Text)

Undocumented member.

improvementSummary_questionId :: Lens' ImprovementSummary (Maybe Text)

Undocumented member.