Amazonka.WellArchitected.Types.Workload

Description

Documentation

data Workload

A workload return object.

See: newWorkload smart constructor.

Constructors

Workload'

Fields

isReviewOwnerUpdateAcknowledged :: Maybe Bool

Flag indicating whether the workload owner has acknowledged that the Review owner field is required.

If a Review owner is not added to the workload within 60 days of acknowledgement, access to the workload is restricted until an owner is added.

architecturalDesign :: Maybe Text

accountIds :: Maybe [Text]

lenses :: Maybe [Text]

reviewRestrictionDate :: Maybe POSIX

industry :: Maybe Text

environment :: Maybe WorkloadEnvironment

riskCounts :: Maybe (HashMap Risk Natural)

awsRegions :: Maybe [Text]

owner :: Maybe Text

improvementStatus :: Maybe WorkloadImprovementStatus

workloadArn :: Maybe Text

industryType :: Maybe Text

workloadId :: Maybe Text

workloadName :: Maybe Text

updatedAt :: Maybe POSIX

notes :: Maybe Text

reviewOwner :: Maybe Text

description :: Maybe Text

pillarPriorities :: Maybe [Text]

shareInvitationId :: Maybe Text

The ID assigned to the share invitation.

nonAwsRegions :: Maybe [Text]

tags :: Maybe (HashMap Text Text)

The tags associated with the workload.

newWorkload :: Workload

Create a value of Workload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isReviewOwnerUpdateAcknowledged:Workload', workload_isReviewOwnerUpdateAcknowledged - Flag indicating whether the workload owner has acknowledged that the Review owner field is required.

If a Review owner is not added to the workload within 60 days of acknowledgement, access to the workload is restricted until an owner is added.

$sel:architecturalDesign:Workload', workload_architecturalDesign - Undocumented member.

$sel:accountIds:Workload', workload_accountIds - Undocumented member.

$sel:lenses:Workload', workload_lenses - Undocumented member.

$sel:reviewRestrictionDate:Workload', workload_reviewRestrictionDate - Undocumented member.

$sel:industry:Workload', workload_industry - Undocumented member.

$sel:environment:Workload', workload_environment - Undocumented member.

$sel:riskCounts:Workload', workload_riskCounts - Undocumented member.

$sel:awsRegions:Workload', workload_awsRegions - Undocumented member.

$sel:owner:Workload', workload_owner - Undocumented member.

$sel:improvementStatus:Workload', workload_improvementStatus - Undocumented member.

$sel:workloadArn:Workload', workload_workloadArn - Undocumented member.

$sel:industryType:Workload', workload_industryType - Undocumented member.

$sel:workloadId:Workload', workload_workloadId - Undocumented member.

$sel:workloadName:Workload', workload_workloadName - Undocumented member.

$sel:updatedAt:Workload', workload_updatedAt - Undocumented member.

$sel:notes:Workload', workload_notes - Undocumented member.

$sel:reviewOwner:Workload', workload_reviewOwner - Undocumented member.

$sel:description:Workload', workload_description - Undocumented member.

$sel:pillarPriorities:Workload', workload_pillarPriorities - Undocumented member.

$sel:shareInvitationId:Workload', workload_shareInvitationId - The ID assigned to the share invitation.

$sel:nonAwsRegions:Workload', workload_nonAwsRegions - Undocumented member.

$sel:tags:Workload', workload_tags - The tags associated with the workload.

workload_isReviewOwnerUpdateAcknowledged :: Lens' Workload (Maybe Bool)

Flag indicating whether the workload owner has acknowledged that the Review owner field is required.

If a Review owner is not added to the workload within 60 days of acknowledgement, access to the workload is restricted until an owner is added.

workload_architecturalDesign :: Lens' Workload (Maybe Text)

Undocumented member.

workload_accountIds :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_lenses :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_reviewRestrictionDate :: Lens' Workload (Maybe UTCTime)

Undocumented member.

workload_industry :: Lens' Workload (Maybe Text)

Undocumented member.

workload_environment :: Lens' Workload (Maybe WorkloadEnvironment)

Undocumented member.

workload_riskCounts :: Lens' Workload (Maybe (HashMap Risk Natural))

Undocumented member.

workload_awsRegions :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_owner :: Lens' Workload (Maybe Text)

Undocumented member.

workload_improvementStatus :: Lens' Workload (Maybe WorkloadImprovementStatus)

Undocumented member.

workload_workloadArn :: Lens' Workload (Maybe Text)

Undocumented member.

workload_industryType :: Lens' Workload (Maybe Text)

Undocumented member.

workload_workloadId :: Lens' Workload (Maybe Text)

Undocumented member.

workload_workloadName :: Lens' Workload (Maybe Text)

Undocumented member.

workload_updatedAt :: Lens' Workload (Maybe UTCTime)

Undocumented member.

workload_notes :: Lens' Workload (Maybe Text)

Undocumented member.

workload_reviewOwner :: Lens' Workload (Maybe Text)

Undocumented member.

workload_description :: Lens' Workload (Maybe Text)

Undocumented member.

workload_pillarPriorities :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_shareInvitationId :: Lens' Workload (Maybe Text)

The ID assigned to the share invitation.

workload_nonAwsRegions :: Lens' Workload (Maybe [Text])

Undocumented member.

workload_tags :: Lens' Workload (Maybe (HashMap Text Text))

The tags associated with the workload.