Amazonka.WellArchitected.UpdateWorkloadShare

Description

Update a workload share.

Creating a Request

data UpdateWorkloadShare

Input for Update Workload Share

See: newUpdateWorkloadShare smart constructor.

Constructors

UpdateWorkloadShare'

Fields

shareId :: Text

workloadId :: Text

permissionType :: PermissionType

newUpdateWorkloadShare

Arguments

:: Text
$sel:shareId:UpdateWorkloadShare'

-> Text
$sel:workloadId:UpdateWorkloadShare'

-> PermissionType
$sel:permissionType:UpdateWorkloadShare'

-> UpdateWorkloadShare

Create a value of UpdateWorkloadShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shareId:UpdateWorkloadShare', updateWorkloadShare_shareId - Undocumented member.

$sel:workloadId:UpdateWorkloadShare', updateWorkloadShare_workloadId - Undocumented member.

$sel:permissionType:UpdateWorkloadShare', updateWorkloadShare_permissionType - Undocumented member.

Request Lenses

updateWorkloadShare_shareId :: Lens' UpdateWorkloadShare Text

Undocumented member.

updateWorkloadShare_workloadId :: Lens' UpdateWorkloadShare Text

Undocumented member.

updateWorkloadShare_permissionType :: Lens' UpdateWorkloadShare PermissionType

Undocumented member.

Destructuring the Response

data UpdateWorkloadShareResponse

Input for Update Workload Share

See: newUpdateWorkloadShareResponse smart constructor.

Constructors

UpdateWorkloadShareResponse'

Fields

workloadShare :: Maybe WorkloadShare

workloadId :: Maybe Text

httpStatus :: Int

The response's http status code.

newUpdateWorkloadShareResponse

Arguments

:: Int
$sel:httpStatus:UpdateWorkloadShareResponse'

-> UpdateWorkloadShareResponse

Create a value of UpdateWorkloadShareResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workloadShare:UpdateWorkloadShareResponse', updateWorkloadShareResponse_workloadShare - Undocumented member.

$sel:workloadId:UpdateWorkloadShare', updateWorkloadShareResponse_workloadId - Undocumented member.

$sel:httpStatus:UpdateWorkloadShareResponse', updateWorkloadShareResponse_httpStatus - The response's http status code.

Response Lenses

updateWorkloadShareResponse_workloadShare :: Lens' UpdateWorkloadShareResponse (Maybe WorkloadShare)

Undocumented member.

updateWorkloadShareResponse_workloadId :: Lens' UpdateWorkloadShareResponse (Maybe Text)

Undocumented member.

updateWorkloadShareResponse_httpStatus :: Lens' UpdateWorkloadShareResponse Int

The response's http status code.