Amazonka.Wisdom.CreateAssistantAssociation

Description

Creates an association between an Amazon Connect Wisdom assistant and another resource. Currently, the only supported association is with a knowledge base. An assistant can have only a single association.

Creating a Request

data CreateAssistantAssociation

See: newCreateAssistantAssociation smart constructor.

Constructors

CreateAssistantAssociation'

Fields

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

assistantId :: Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

association :: AssistantAssociationInputData

The identifier of the associated resource.

associationType :: AssociationType

The type of association.

newCreateAssistantAssociation

Arguments

:: Text
$sel:assistantId:CreateAssistantAssociation'

-> AssistantAssociationInputData
$sel:association:CreateAssistantAssociation'

-> AssociationType
$sel:associationType:CreateAssistantAssociation'

-> CreateAssistantAssociation

Create a value of CreateAssistantAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateAssistantAssociation', createAssistantAssociation_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:tags:CreateAssistantAssociation', createAssistantAssociation_tags - The tags used to organize, track, or control access for this resource.

$sel:assistantId:CreateAssistantAssociation', createAssistantAssociation_assistantId - The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

$sel:association:CreateAssistantAssociation', createAssistantAssociation_association - The identifier of the associated resource.

$sel:associationType:CreateAssistantAssociation', createAssistantAssociation_associationType - The type of association.

Request Lenses

createAssistantAssociation_clientToken :: Lens' CreateAssistantAssociation (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createAssistantAssociation_tags :: Lens' CreateAssistantAssociation (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createAssistantAssociation_assistantId :: Lens' CreateAssistantAssociation Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

createAssistantAssociation_association :: Lens' CreateAssistantAssociation AssistantAssociationInputData

The identifier of the associated resource.

createAssistantAssociation_associationType :: Lens' CreateAssistantAssociation AssociationType

The type of association.

Destructuring the Response

data CreateAssistantAssociationResponse

See: newCreateAssistantAssociationResponse smart constructor.

Constructors

CreateAssistantAssociationResponse'

Fields

assistantAssociation :: Maybe AssistantAssociationData

The assistant association.

httpStatus :: Int

The response's http status code.

newCreateAssistantAssociationResponse

Arguments

:: Int
$sel:httpStatus:CreateAssistantAssociationResponse'

-> CreateAssistantAssociationResponse

Create a value of CreateAssistantAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assistantAssociation:CreateAssistantAssociationResponse', createAssistantAssociationResponse_assistantAssociation - The assistant association.

$sel:httpStatus:CreateAssistantAssociationResponse', createAssistantAssociationResponse_httpStatus - The response's http status code.

Response Lenses

createAssistantAssociationResponse_assistantAssociation :: Lens' CreateAssistantAssociationResponse (Maybe AssistantAssociationData)

The assistant association.

createAssistantAssociationResponse_httpStatus :: Lens' CreateAssistantAssociationResponse Int

The response's http status code.