Amazonka.Wisdom.CreateSession

Description

Creates a session. A session is a contextual container used for generating recommendations. Amazon Connect creates a new Wisdom session for each contact on which Wisdom is enabled.

Creating a Request

data CreateSession

See: newCreateSession smart constructor.

Constructors

CreateSession'

Fields

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

description :: Maybe Text

The description.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

assistantId :: Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

name :: Text

The name of the session.

newCreateSession

Arguments

:: Text
$sel:assistantId:CreateSession'

-> Text
$sel:name:CreateSession'

-> CreateSession

Create a value of CreateSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateSession', createSession_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:description:CreateSession', createSession_description - The description.

$sel:tags:CreateSession', createSession_tags - The tags used to organize, track, or control access for this resource.

$sel:assistantId:CreateSession', createSession_assistantId - The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

$sel:name:CreateSession', createSession_name - The name of the session.

Request Lenses

createSession_clientToken :: Lens' CreateSession (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createSession_description :: Lens' CreateSession (Maybe Text)

The description.

createSession_tags :: Lens' CreateSession (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createSession_assistantId :: Lens' CreateSession Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

createSession_name :: Lens' CreateSession Text

The name of the session.

Destructuring the Response

data CreateSessionResponse

See: newCreateSessionResponse smart constructor.

Constructors

CreateSessionResponse'

Fields

session :: Maybe SessionData

The session.

httpStatus :: Int

The response's http status code.

newCreateSessionResponse

Arguments

:: Int
$sel:httpStatus:CreateSessionResponse'

-> CreateSessionResponse

Create a value of CreateSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:session:CreateSessionResponse', createSessionResponse_session - The session.

$sel:httpStatus:CreateSessionResponse', createSessionResponse_httpStatus - The response's http status code.

Response Lenses

createSessionResponse_session :: Lens' CreateSessionResponse (Maybe SessionData)

The session.

createSessionResponse_httpStatus :: Lens' CreateSessionResponse Int

The response's http status code.