Amazonka.Wisdom.GetAssistantAssociation

Description

Retrieves information about an assistant association.

Creating a Request

data GetAssistantAssociation

See: newGetAssistantAssociation smart constructor.

Constructors

GetAssistantAssociation'

Fields

assistantAssociationId :: Text

The identifier of the assistant association. Can be either the ID or the ARN. URLs cannot contain the ARN.

assistantId :: Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

newGetAssistantAssociation

Arguments

:: Text
$sel:assistantAssociationId:GetAssistantAssociation'

-> Text
$sel:assistantId:GetAssistantAssociation'

-> GetAssistantAssociation

Create a value of GetAssistantAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assistantAssociationId:GetAssistantAssociation', getAssistantAssociation_assistantAssociationId - The identifier of the assistant association. Can be either the ID or the ARN. URLs cannot contain the ARN.

$sel:assistantId:GetAssistantAssociation', getAssistantAssociation_assistantId - The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

Request Lenses

getAssistantAssociation_assistantAssociationId :: Lens' GetAssistantAssociation Text

The identifier of the assistant association. Can be either the ID or the ARN. URLs cannot contain the ARN.

getAssistantAssociation_assistantId :: Lens' GetAssistantAssociation Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

Destructuring the Response

data GetAssistantAssociationResponse

See: newGetAssistantAssociationResponse smart constructor.

Constructors

GetAssistantAssociationResponse'

Fields

assistantAssociation :: Maybe AssistantAssociationData

The assistant association.

httpStatus :: Int

The response's http status code.

newGetAssistantAssociationResponse

Arguments

:: Int
$sel:httpStatus:GetAssistantAssociationResponse'

-> GetAssistantAssociationResponse

Create a value of GetAssistantAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assistantAssociation:GetAssistantAssociationResponse', getAssistantAssociationResponse_assistantAssociation - The assistant association.

$sel:httpStatus:GetAssistantAssociationResponse', getAssistantAssociationResponse_httpStatus - The response's http status code.

Response Lenses

getAssistantAssociationResponse_assistantAssociation :: Lens' GetAssistantAssociationResponse (Maybe AssistantAssociationData)

The assistant association.

getAssistantAssociationResponse_httpStatus :: Lens' GetAssistantAssociationResponse Int

The response's http status code.