Amazonka.Wisdom.GetKnowledgeBase

Description

Retrieves information about the knowledge base.

Creating a Request

data GetKnowledgeBase

See: newGetKnowledgeBase smart constructor.

Constructors

GetKnowledgeBase'

Fields

knowledgeBaseId :: Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

newGetKnowledgeBase

Arguments

:: Text
$sel:knowledgeBaseId:GetKnowledgeBase'

-> GetKnowledgeBase

Create a value of GetKnowledgeBase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:knowledgeBaseId:GetKnowledgeBase', getKnowledgeBase_knowledgeBaseId - The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Request Lenses

getKnowledgeBase_knowledgeBaseId :: Lens' GetKnowledgeBase Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Destructuring the Response

data GetKnowledgeBaseResponse

See: newGetKnowledgeBaseResponse smart constructor.

Constructors

GetKnowledgeBaseResponse'

Fields

knowledgeBase :: Maybe KnowledgeBaseData

The knowledge base.

httpStatus :: Int

The response's http status code.

newGetKnowledgeBaseResponse

Arguments

:: Int
$sel:httpStatus:GetKnowledgeBaseResponse'

-> GetKnowledgeBaseResponse

Create a value of GetKnowledgeBaseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:knowledgeBase:GetKnowledgeBaseResponse', getKnowledgeBaseResponse_knowledgeBase - The knowledge base.

$sel:httpStatus:GetKnowledgeBaseResponse', getKnowledgeBaseResponse_httpStatus - The response's http status code.

Response Lenses

getKnowledgeBaseResponse_knowledgeBase :: Lens' GetKnowledgeBaseResponse (Maybe KnowledgeBaseData)

The knowledge base.

getKnowledgeBaseResponse_httpStatus :: Lens' GetKnowledgeBaseResponse Int

The response's http status code.