Amazonka.Wisdom.GetSession

Description

Retrieves information for a specified session.

Creating a Request

data GetSession

See: newGetSession smart constructor.

Constructors

GetSession'

Fields

assistantId :: Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

sessionId :: Text

The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

newGetSession

Arguments

:: Text
$sel:assistantId:GetSession'

-> Text
$sel:sessionId:GetSession'

-> GetSession

Create a value of GetSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assistantId:GetSession', getSession_assistantId - The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

$sel:sessionId:GetSession', getSession_sessionId - The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

Request Lenses

getSession_assistantId :: Lens' GetSession Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

getSession_sessionId :: Lens' GetSession Text

The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

Destructuring the Response

data GetSessionResponse

See: newGetSessionResponse smart constructor.

Constructors

GetSessionResponse'

Fields

session :: Maybe SessionData

The session.

httpStatus :: Int

The response's http status code.

newGetSessionResponse

Arguments

:: Int
$sel:httpStatus:GetSessionResponse'

-> GetSessionResponse

Create a value of GetSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:session:GetSessionResponse', getSessionResponse_session - The session.

$sel:httpStatus:GetSessionResponse', getSessionResponse_httpStatus - The response's http status code.

Response Lenses

getSessionResponse_session :: Lens' GetSessionResponse (Maybe SessionData)

The session.

getSessionResponse_httpStatus :: Lens' GetSessionResponse Int

The response's http status code.