Amazonka.Wisdom.Lens

Description

Operations

ListAssistantAssociations

listAssistantAssociations_nextToken :: Lens' ListAssistantAssociations (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listAssistantAssociations_maxResults :: Lens' ListAssistantAssociations (Maybe Natural)

The maximum number of results to return per page.

listAssistantAssociations_assistantId :: Lens' ListAssistantAssociations Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

listAssistantAssociationsResponse_nextToken :: Lens' ListAssistantAssociationsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listAssistantAssociationsResponse_httpStatus :: Lens' ListAssistantAssociationsResponse Int

The response's http status code.

listAssistantAssociationsResponse_assistantAssociationSummaries :: Lens' ListAssistantAssociationsResponse [AssistantAssociationSummary]

Summary information about assistant associations.

GetRecommendations

getRecommendations_waitTimeSeconds :: Lens' GetRecommendations (Maybe Natural)

The duration (in seconds) for which the call waits for a recommendation to be made available before returning. If a recommendation is available, the call returns sooner than WaitTimeSeconds. If no messages are available and the wait time expires, the call returns successfully with an empty list.

getRecommendations_maxResults :: Lens' GetRecommendations (Maybe Natural)

The maximum number of results to return per page.

getRecommendations_assistantId :: Lens' GetRecommendations Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

getRecommendations_sessionId :: Lens' GetRecommendations Text

The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

getRecommendationsResponse_httpStatus :: Lens' GetRecommendationsResponse Int

The response's http status code.

getRecommendationsResponse_recommendations :: Lens' GetRecommendationsResponse [RecommendationData]

The recommendations.

SearchContent

searchContent_nextToken :: Lens' SearchContent (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

searchContent_maxResults :: Lens' SearchContent (Maybe Natural)

The maximum number of results to return per page.

searchContent_knowledgeBaseId :: Lens' SearchContent Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

searchContent_searchExpression :: Lens' SearchContent SearchExpression

The search expression to filter results.

searchContentResponse_nextToken :: Lens' SearchContentResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

searchContentResponse_httpStatus :: Lens' SearchContentResponse Int

The response's http status code.

searchContentResponse_contentSummaries :: Lens' SearchContentResponse [ContentSummary]

Summary information about the content.

RemoveKnowledgeBaseTemplateUri

removeKnowledgeBaseTemplateUri_knowledgeBaseId :: Lens' RemoveKnowledgeBaseTemplateUri Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

removeKnowledgeBaseTemplateUriResponse_httpStatus :: Lens' RemoveKnowledgeBaseTemplateUriResponse Int

The response's http status code.

GetAssistant

getAssistant_assistantId :: Lens' GetAssistant Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

getAssistantResponse_assistant :: Lens' GetAssistantResponse (Maybe AssistantData)

Information about the assistant.

getAssistantResponse_httpStatus :: Lens' GetAssistantResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

ListAssistants

listAssistants_nextToken :: Lens' ListAssistants (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listAssistants_maxResults :: Lens' ListAssistants (Maybe Natural)

The maximum number of results to return per page.

listAssistantsResponse_nextToken :: Lens' ListAssistantsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listAssistantsResponse_httpStatus :: Lens' ListAssistantsResponse Int

The response's http status code.

listAssistantsResponse_assistantSummaries :: Lens' ListAssistantsResponse [AssistantSummary]

Information about the assistants.

UpdateKnowledgeBaseTemplateUri

updateKnowledgeBaseTemplateUri_knowledgeBaseId :: Lens' UpdateKnowledgeBaseTemplateUri Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

updateKnowledgeBaseTemplateUri_templateUri :: Lens' UpdateKnowledgeBaseTemplateUri Text

The template URI to update.

updateKnowledgeBaseTemplateUriResponse_knowledgeBase :: Lens' UpdateKnowledgeBaseTemplateUriResponse (Maybe KnowledgeBaseData)

The knowledge base to update.

updateKnowledgeBaseTemplateUriResponse_httpStatus :: Lens' UpdateKnowledgeBaseTemplateUriResponse Int

The response's http status code.

ListContents

listContents_nextToken :: Lens' ListContents (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listContents_maxResults :: Lens' ListContents (Maybe Natural)

The maximum number of results to return per page.

listContents_knowledgeBaseId :: Lens' ListContents Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

listContentsResponse_nextToken :: Lens' ListContentsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listContentsResponse_httpStatus :: Lens' ListContentsResponse Int

The response's http status code.

listContentsResponse_contentSummaries :: Lens' ListContentsResponse [ContentSummary]

Information about the content.

DeleteContent

deleteContent_contentId :: Lens' DeleteContent Text

The identifier of the content. Can be either the ID or the ARN. URLs cannot contain the ARN.

deleteContent_knowledgeBaseId :: Lens' DeleteContent Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

deleteContentResponse_httpStatus :: Lens' DeleteContentResponse Int

The response's http status code.

UpdateContent

updateContent_overrideLinkOutUri :: Lens' UpdateContent (Maybe Text)

The URI for the article. If the knowledge base has a templateUri, setting this argument overrides it for this piece of content. To remove an existing overrideLinkOurUri, exclude this argument and set removeOverrideLinkOutUri to true.

updateContent_removeOverrideLinkOutUri :: Lens' UpdateContent (Maybe Bool)

Unset the existing overrideLinkOutUri if it exists.

updateContent_metadata :: Lens' UpdateContent (Maybe (HashMap Text Text))

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

updateContent_title :: Lens' UpdateContent (Maybe Text)

The title of the content.

updateContent_revisionId :: Lens' UpdateContent (Maybe Text)

The revisionId of the content resource to update, taken from an earlier call to GetContent, GetContentSummary, SearchContent, or ListContents. If included, this argument acts as an optimistic lock to ensure content was not modified since it was last read. If it has been modified, this API throws a PreconditionFailedException.

updateContent_uploadId :: Lens' UpdateContent (Maybe Text)

A pointer to the uploaded asset. This value is returned by StartContentUpload.

updateContent_contentId :: Lens' UpdateContent Text

The identifier of the content. Can be either the ID or the ARN. URLs cannot contain the ARN.

updateContent_knowledgeBaseId :: Lens' UpdateContent Text

The the identifier of the knowledge base. Can be either the ID or the ARN

updateContentResponse_content :: Lens' UpdateContentResponse (Maybe ContentData)

The content.

updateContentResponse_httpStatus :: Lens' UpdateContentResponse Int

The response's http status code.

CreateAssistant

createAssistant_clientToken :: Lens' CreateAssistant (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createAssistant_description :: Lens' CreateAssistant (Maybe Text)

The description of the assistant.

createAssistant_serverSideEncryptionConfiguration :: Lens' CreateAssistant (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

createAssistant_tags :: Lens' CreateAssistant (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createAssistant_name :: Lens' CreateAssistant Text

The name of the assistant.

createAssistant_type :: Lens' CreateAssistant AssistantType

The type of assistant.

createAssistantResponse_assistant :: Lens' CreateAssistantResponse (Maybe AssistantData)

Information about the assistant.

createAssistantResponse_httpStatus :: Lens' CreateAssistantResponse Int

The response's http status code.

GetContentSummary

getContentSummary_contentId :: Lens' GetContentSummary Text

The identifier of the content. Can be either the ID or the ARN. URLs cannot contain the ARN.

getContentSummary_knowledgeBaseId :: Lens' GetContentSummary Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

getContentSummaryResponse_contentSummary :: Lens' GetContentSummaryResponse (Maybe ContentSummary)

The content summary.

getContentSummaryResponse_httpStatus :: Lens' GetContentSummaryResponse Int

The response's http status code.

NotifyRecommendationsReceived

notifyRecommendationsReceived_assistantId :: Lens' NotifyRecommendationsReceived Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

notifyRecommendationsReceived_recommendationIds :: Lens' NotifyRecommendationsReceived [Text]

The identifiers of the recommendations.

notifyRecommendationsReceived_sessionId :: Lens' NotifyRecommendationsReceived Text

The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

notifyRecommendationsReceivedResponse_recommendationIds :: Lens' NotifyRecommendationsReceivedResponse (Maybe [Text])

The identifiers of the recommendations.

notifyRecommendationsReceivedResponse_errors :: Lens' NotifyRecommendationsReceivedResponse (Maybe [NotifyRecommendationsReceivedError])

The identifiers of recommendations that are causing errors.

notifyRecommendationsReceivedResponse_httpStatus :: Lens' NotifyRecommendationsReceivedResponse Int

The response's http status code.

DeleteAssistantAssociation

deleteAssistantAssociation_assistantAssociationId :: Lens' DeleteAssistantAssociation Text

The identifier of the assistant association. Can be either the ID or the ARN. URLs cannot contain the ARN.

deleteAssistantAssociation_assistantId :: Lens' DeleteAssistantAssociation Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

deleteAssistantAssociationResponse_httpStatus :: Lens' DeleteAssistantAssociationResponse Int

The response's http status code.

GetContent

getContent_contentId :: Lens' GetContent Text

The identifier of the content. Can be either the ID or the ARN. URLs cannot contain the ARN.

getContent_knowledgeBaseId :: Lens' GetContent Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

getContentResponse_content :: Lens' GetContentResponse (Maybe ContentData)

The content.

getContentResponse_httpStatus :: Lens' GetContentResponse Int

The response's http status code.

StartContentUpload

startContentUpload_contentType :: Lens' StartContentUpload Text

The type of content to upload.

startContentUpload_knowledgeBaseId :: Lens' StartContentUpload Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

startContentUploadResponse_httpStatus :: Lens' StartContentUploadResponse Int

The response's http status code.

startContentUploadResponse_headersToInclude :: Lens' StartContentUploadResponse (HashMap Text Text)

The headers to include in the upload.

startContentUploadResponse_uploadId :: Lens' StartContentUploadResponse Text

The identifier of the upload.

startContentUploadResponse_url :: Lens' StartContentUploadResponse Text

The URL of the upload.

startContentUploadResponse_urlExpiry :: Lens' StartContentUploadResponse UTCTime

The expiration time of the URL as an epoch timestamp.

CreateSession

createSession_clientToken :: Lens' CreateSession (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createSession_description :: Lens' CreateSession (Maybe Text)

The description.

createSession_tags :: Lens' CreateSession (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createSession_assistantId :: Lens' CreateSession Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

createSession_name :: Lens' CreateSession Text

The name of the session.

createSessionResponse_session :: Lens' CreateSessionResponse (Maybe SessionData)

The session.

createSessionResponse_httpStatus :: Lens' CreateSessionResponse Int

The response's http status code.

CreateContent

createContent_overrideLinkOutUri :: Lens' CreateContent (Maybe Text)

The URI you want to use for the article. If the knowledge base has a templateUri, setting this argument overrides it for this piece of content.

createContent_clientToken :: Lens' CreateContent (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createContent_metadata :: Lens' CreateContent (Maybe (HashMap Text Text))

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

createContent_title :: Lens' CreateContent (Maybe Text)

The title of the content. If not set, the title is equal to the name.

createContent_tags :: Lens' CreateContent (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createContent_knowledgeBaseId :: Lens' CreateContent Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

createContent_name :: Lens' CreateContent Text

The name of the content. Each piece of content in a knowledge base must have a unique name. You can retrieve a piece of content using only its knowledge base and its name with the SearchContent API.

createContent_uploadId :: Lens' CreateContent Text

A pointer to the uploaded asset. This value is returned by StartContentUpload.

createContentResponse_content :: Lens' CreateContentResponse (Maybe ContentData)

The content.

createContentResponse_httpStatus :: Lens' CreateContentResponse Int

The response's http status code.

DeleteAssistant

deleteAssistant_assistantId :: Lens' DeleteAssistant Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

deleteAssistantResponse_httpStatus :: Lens' DeleteAssistantResponse Int

The response's http status code.

GetSession

getSession_assistantId :: Lens' GetSession Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

getSession_sessionId :: Lens' GetSession Text

The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

getSessionResponse_session :: Lens' GetSessionResponse (Maybe SessionData)

The session.

getSessionResponse_httpStatus :: Lens' GetSessionResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

CreateKnowledgeBase

createKnowledgeBase_clientToken :: Lens' CreateKnowledgeBase (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createKnowledgeBase_renderingConfiguration :: Lens' CreateKnowledgeBase (Maybe RenderingConfiguration)

Information about how to render the content.

createKnowledgeBase_sourceConfiguration :: Lens' CreateKnowledgeBase (Maybe SourceConfiguration)

The source of the knowledge base content. Only set this argument for EXTERNAL knowledge bases.

createKnowledgeBase_description :: Lens' CreateKnowledgeBase (Maybe Text)

The description.

createKnowledgeBase_serverSideEncryptionConfiguration :: Lens' CreateKnowledgeBase (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

createKnowledgeBase_tags :: Lens' CreateKnowledgeBase (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createKnowledgeBase_knowledgeBaseType :: Lens' CreateKnowledgeBase KnowledgeBaseType

The type of knowledge base. Only CUSTOM knowledge bases allow you to upload your own content. EXTERNAL knowledge bases support integrations with third-party systems whose content is synchronized automatically.

createKnowledgeBase_name :: Lens' CreateKnowledgeBase Text

The name of the knowledge base.

createKnowledgeBaseResponse_knowledgeBase :: Lens' CreateKnowledgeBaseResponse (Maybe KnowledgeBaseData)

The knowledge base.

createKnowledgeBaseResponse_httpStatus :: Lens' CreateKnowledgeBaseResponse Int

The response's http status code.

GetAssistantAssociation

getAssistantAssociation_assistantAssociationId :: Lens' GetAssistantAssociation Text

The identifier of the assistant association. Can be either the ID or the ARN. URLs cannot contain the ARN.

getAssistantAssociation_assistantId :: Lens' GetAssistantAssociation Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

getAssistantAssociationResponse_assistantAssociation :: Lens' GetAssistantAssociationResponse (Maybe AssistantAssociationData)

The assistant association.

getAssistantAssociationResponse_httpStatus :: Lens' GetAssistantAssociationResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The tag keys.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

SearchSessions

searchSessions_nextToken :: Lens' SearchSessions (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

searchSessions_maxResults :: Lens' SearchSessions (Maybe Natural)

The maximum number of results to return per page.

searchSessions_assistantId :: Lens' SearchSessions Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

searchSessions_searchExpression :: Lens' SearchSessions SearchExpression

The search expression to filter results.

searchSessionsResponse_nextToken :: Lens' SearchSessionsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

searchSessionsResponse_httpStatus :: Lens' SearchSessionsResponse Int

The response's http status code.

searchSessionsResponse_sessionSummaries :: Lens' SearchSessionsResponse [SessionSummary]

Summary information about the sessions.

ListKnowledgeBases

listKnowledgeBases_nextToken :: Lens' ListKnowledgeBases (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listKnowledgeBases_maxResults :: Lens' ListKnowledgeBases (Maybe Natural)

The maximum number of results to return per page.

listKnowledgeBasesResponse_nextToken :: Lens' ListKnowledgeBasesResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listKnowledgeBasesResponse_httpStatus :: Lens' ListKnowledgeBasesResponse Int

The response's http status code.

listKnowledgeBasesResponse_knowledgeBaseSummaries :: Lens' ListKnowledgeBasesResponse [KnowledgeBaseSummary]

Information about the knowledge bases.

QueryAssistant

queryAssistant_nextToken :: Lens' QueryAssistant (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

queryAssistant_maxResults :: Lens' QueryAssistant (Maybe Natural)

The maximum number of results to return per page.

queryAssistant_assistantId :: Lens' QueryAssistant Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

queryAssistant_queryText :: Lens' QueryAssistant Text

The text to search for.

queryAssistantResponse_nextToken :: Lens' QueryAssistantResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

queryAssistantResponse_httpStatus :: Lens' QueryAssistantResponse Int

The response's http status code.

queryAssistantResponse_results :: Lens' QueryAssistantResponse [ResultData]

The results of the query.

DeleteKnowledgeBase

deleteKnowledgeBase_knowledgeBaseId :: Lens' DeleteKnowledgeBase Text

The knowledge base to delete content from. Can be either the ID or the ARN. URLs cannot contain the ARN.

deleteKnowledgeBaseResponse_httpStatus :: Lens' DeleteKnowledgeBaseResponse Int

The response's http status code.

CreateAssistantAssociation

createAssistantAssociation_clientToken :: Lens' CreateAssistantAssociation (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createAssistantAssociation_tags :: Lens' CreateAssistantAssociation (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

createAssistantAssociation_assistantId :: Lens' CreateAssistantAssociation Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

createAssistantAssociation_association :: Lens' CreateAssistantAssociation AssistantAssociationInputData

The identifier of the associated resource.

createAssistantAssociation_associationType :: Lens' CreateAssistantAssociation AssociationType

The type of association.

createAssistantAssociationResponse_assistantAssociation :: Lens' CreateAssistantAssociationResponse (Maybe AssistantAssociationData)

The assistant association.

createAssistantAssociationResponse_httpStatus :: Lens' CreateAssistantAssociationResponse Int

The response's http status code.

GetKnowledgeBase

getKnowledgeBase_knowledgeBaseId :: Lens' GetKnowledgeBase Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

getKnowledgeBaseResponse_knowledgeBase :: Lens' GetKnowledgeBaseResponse (Maybe KnowledgeBaseData)

The knowledge base.

getKnowledgeBaseResponse_httpStatus :: Lens' GetKnowledgeBaseResponse Int

The response's http status code.

Types

AppIntegrationsConfiguration

appIntegrationsConfiguration_appIntegrationArn :: Lens' AppIntegrationsConfiguration Text

The Amazon Resource Name (ARN) of the AppIntegrations DataIntegration to use for ingesting content.

appIntegrationsConfiguration_objectFields :: Lens' AppIntegrationsConfiguration (NonEmpty Text)

The fields from the source that are made available to your agents in Wisdom.

For Salesforce, you must include at least Id, ArticleNumber, VersionNumber, Title, PublishStatus, and IsDeleted.

For ServiceNow, you must include at least number, short_description, sys_mod_count, workflow_state, and active.

Make sure to include additional field(s); these are indexed and used to source recommendations.

AssistantAssociationData

assistantAssociationData_tags :: Lens' AssistantAssociationData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

assistantAssociationData_assistantArn :: Lens' AssistantAssociationData Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantAssociationData_assistantAssociationArn :: Lens' AssistantAssociationData Text

The Amazon Resource Name (ARN) of the assistant association.

assistantAssociationData_assistantAssociationId :: Lens' AssistantAssociationData Text

The identifier of the assistant association.

assistantAssociationData_assistantId :: Lens' AssistantAssociationData Text

The identifier of the Wisdom assistant.

assistantAssociationData_associationData :: Lens' AssistantAssociationData AssistantAssociationOutputData

A union type that currently has a single argument, the knowledge base ID.

assistantAssociationData_associationType :: Lens' AssistantAssociationData AssociationType

The type of association.

AssistantAssociationInputData

assistantAssociationInputData_knowledgeBaseId :: Lens' AssistantAssociationInputData (Maybe Text)

The the identifier of the knowledge base.

AssistantAssociationOutputData

assistantAssociationOutputData_knowledgeBaseAssociation :: Lens' AssistantAssociationOutputData (Maybe KnowledgeBaseAssociationData)

The knowledge base where output data is sent.

AssistantAssociationSummary

assistantAssociationSummary_tags :: Lens' AssistantAssociationSummary (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

assistantAssociationSummary_assistantArn :: Lens' AssistantAssociationSummary Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantAssociationSummary_assistantAssociationArn :: Lens' AssistantAssociationSummary Text

The Amazon Resource Name (ARN) of the assistant association.

assistantAssociationSummary_assistantAssociationId :: Lens' AssistantAssociationSummary Text

The identifier of the assistant association.

assistantAssociationSummary_assistantId :: Lens' AssistantAssociationSummary Text

The identifier of the Wisdom assistant.

assistantAssociationSummary_associationData :: Lens' AssistantAssociationSummary AssistantAssociationOutputData

The association data.

assistantAssociationSummary_associationType :: Lens' AssistantAssociationSummary AssociationType

The type of association.

AssistantData

assistantData_description :: Lens' AssistantData (Maybe Text)

The description.

assistantData_serverSideEncryptionConfiguration :: Lens' AssistantData (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

assistantData_tags :: Lens' AssistantData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

assistantData_assistantArn :: Lens' AssistantData Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantData_assistantId :: Lens' AssistantData Text

The identifier of the Wisdom assistant.

assistantData_name :: Lens' AssistantData Text

The name.

assistantData_status :: Lens' AssistantData AssistantStatus

The status of the assistant.

assistantData_type :: Lens' AssistantData AssistantType

The type of assistant.

AssistantSummary

assistantSummary_description :: Lens' AssistantSummary (Maybe Text)

The description of the assistant.

assistantSummary_serverSideEncryptionConfiguration :: Lens' AssistantSummary (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

assistantSummary_tags :: Lens' AssistantSummary (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

assistantSummary_assistantArn :: Lens' AssistantSummary Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantSummary_assistantId :: Lens' AssistantSummary Text

The identifier of the Wisdom assistant.

assistantSummary_name :: Lens' AssistantSummary Text

The name of the assistant.

assistantSummary_status :: Lens' AssistantSummary AssistantStatus

The status of the assistant.

assistantSummary_type :: Lens' AssistantSummary AssistantType

The type of the assistant.

ContentData

contentData_linkOutUri :: Lens' ContentData (Maybe Text)

The URI of the content.

contentData_tags :: Lens' ContentData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

contentData_contentArn :: Lens' ContentData Text

The Amazon Resource Name (ARN) of the content.

contentData_contentId :: Lens' ContentData Text

The identifier of the content.

contentData_contentType :: Lens' ContentData Text

The media type of the content.

contentData_knowledgeBaseArn :: Lens' ContentData Text

The Amazon Resource Name (ARN) of the knowledge base.

contentData_knowledgeBaseId :: Lens' ContentData Text

The the identifier of the knowledge base.

contentData_metadata :: Lens' ContentData (HashMap Text Text)

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

contentData_name :: Lens' ContentData Text

The name of the content.

contentData_revisionId :: Lens' ContentData Text

The identifier of the content revision.

contentData_status :: Lens' ContentData ContentStatus

The status of the content.

contentData_title :: Lens' ContentData Text

The title of the content.

contentData_url :: Lens' ContentData Text

The URL of the content.

contentData_urlExpiry :: Lens' ContentData UTCTime

The expiration time of the URL as an epoch timestamp.

ContentReference

contentReference_contentArn :: Lens' ContentReference (Maybe Text)

The Amazon Resource Name (ARN) of the content.

contentReference_contentId :: Lens' ContentReference (Maybe Text)

The identifier of the content.

contentReference_knowledgeBaseArn :: Lens' ContentReference (Maybe Text)

The Amazon Resource Name (ARN) of the knowledge base.

contentReference_knowledgeBaseId :: Lens' ContentReference (Maybe Text)

The the identifier of the knowledge base.

ContentSummary

contentSummary_tags :: Lens' ContentSummary (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

contentSummary_contentArn :: Lens' ContentSummary Text

The Amazon Resource Name (ARN) of the content.

contentSummary_contentId :: Lens' ContentSummary Text

The identifier of the content.

contentSummary_contentType :: Lens' ContentSummary Text

The media type of the content.

contentSummary_knowledgeBaseArn :: Lens' ContentSummary Text

The Amazon Resource Name (ARN) of the knowledge base.

contentSummary_knowledgeBaseId :: Lens' ContentSummary Text

The the identifier of the knowledge base.

contentSummary_metadata :: Lens' ContentSummary (HashMap Text Text)

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

contentSummary_name :: Lens' ContentSummary Text

The name of the content.

contentSummary_revisionId :: Lens' ContentSummary Text

The identifier of the revision of the content.

contentSummary_status :: Lens' ContentSummary ContentStatus

The status of the content.

contentSummary_title :: Lens' ContentSummary Text

The title of the content.

Document

document_excerpt :: Lens' Document (Maybe DocumentText)

The excerpt from the document.

document_title :: Lens' Document (Maybe DocumentText)

The title of the document.

document_contentReference :: Lens' Document ContentReference

A reference to the content resource.

DocumentText

documentText_text :: Lens' DocumentText (Maybe Text)

Text in the document.

documentText_highlights :: Lens' DocumentText (Maybe [Highlight])

Highlights in the document text.

Filter

filter_field :: Lens' Filter FilterField

The field on which to filter.

filter_operator :: Lens' Filter FilterOperator

The operator to use for comparing the field’s value with the provided value.

filter_value :: Lens' Filter Text

The desired field value on which to filter.

Highlight

highlight_endOffsetExclusive :: Lens' Highlight (Maybe Int)

The offset for the end of the highlight.

highlight_beginOffsetInclusive :: Lens' Highlight (Maybe Int)

The offset for the start of the highlight.

KnowledgeBaseAssociationData

knowledgeBaseAssociationData_knowledgeBaseArn :: Lens' KnowledgeBaseAssociationData (Maybe Text)

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseAssociationData_knowledgeBaseId :: Lens' KnowledgeBaseAssociationData (Maybe Text)

The the identifier of the knowledge base.

KnowledgeBaseData

knowledgeBaseData_renderingConfiguration :: Lens' KnowledgeBaseData (Maybe RenderingConfiguration)

Information about how to render the content.

knowledgeBaseData_sourceConfiguration :: Lens' KnowledgeBaseData (Maybe SourceConfiguration)

Source configuration information about the knowledge base.

knowledgeBaseData_lastContentModificationTime :: Lens' KnowledgeBaseData (Maybe UTCTime)

An epoch timestamp indicating the most recent content modification inside the knowledge base. If no content exists in a knowledge base, this value is unset.

knowledgeBaseData_description :: Lens' KnowledgeBaseData (Maybe Text)

The description.

knowledgeBaseData_serverSideEncryptionConfiguration :: Lens' KnowledgeBaseData (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

knowledgeBaseData_tags :: Lens' KnowledgeBaseData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

knowledgeBaseData_knowledgeBaseArn :: Lens' KnowledgeBaseData Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseData_knowledgeBaseId :: Lens' KnowledgeBaseData Text

The the identifier of the knowledge base.

knowledgeBaseData_knowledgeBaseType :: Lens' KnowledgeBaseData KnowledgeBaseType

The type of knowledge base.

knowledgeBaseData_name :: Lens' KnowledgeBaseData Text

The name of the knowledge base.

knowledgeBaseData_status :: Lens' KnowledgeBaseData KnowledgeBaseStatus

The status of the knowledge base.

KnowledgeBaseSummary

knowledgeBaseSummary_renderingConfiguration :: Lens' KnowledgeBaseSummary (Maybe RenderingConfiguration)

Information about how to render the content.

knowledgeBaseSummary_sourceConfiguration :: Lens' KnowledgeBaseSummary (Maybe SourceConfiguration)

KEVIN

knowledgeBaseSummary_description :: Lens' KnowledgeBaseSummary (Maybe Text)

The description of the knowledge base.

knowledgeBaseSummary_serverSideEncryptionConfiguration :: Lens' KnowledgeBaseSummary (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

knowledgeBaseSummary_tags :: Lens' KnowledgeBaseSummary (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

knowledgeBaseSummary_knowledgeBaseArn :: Lens' KnowledgeBaseSummary Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseSummary_knowledgeBaseId :: Lens' KnowledgeBaseSummary Text

The the identifier of the knowledge base.

knowledgeBaseSummary_knowledgeBaseType :: Lens' KnowledgeBaseSummary KnowledgeBaseType

The type of knowledge base.

knowledgeBaseSummary_name :: Lens' KnowledgeBaseSummary Text

The name of the knowledge base.

knowledgeBaseSummary_status :: Lens' KnowledgeBaseSummary KnowledgeBaseStatus

The status of the knowledge base summary.

NotifyRecommendationsReceivedError

notifyRecommendationsReceivedError_recommendationId :: Lens' NotifyRecommendationsReceivedError (Maybe Text)

The identifier of the recommendation that is in error.

notifyRecommendationsReceivedError_message :: Lens' NotifyRecommendationsReceivedError (Maybe Text)

A recommendation is causing an error.

RecommendationData

recommendationData_relevanceLevel :: Lens' RecommendationData (Maybe RelevanceLevel)

The relevance level of the recommendation.

recommendationData_relevanceScore :: Lens' RecommendationData (Maybe Double)

The relevance score of the recommendation.

recommendationData_document :: Lens' RecommendationData Document

The recommended document.

recommendationData_recommendationId :: Lens' RecommendationData Text

The identifier of the recommendation.

RenderingConfiguration

renderingConfiguration_templateUri :: Lens' RenderingConfiguration (Maybe Text)

A URI template containing exactly one variable in ${variableName} format. This can only be set for EXTERNAL knowledge bases. For Salesforce and ServiceNow, the variable must be one of the following:

Salesforce: Id, ArticleNumber, VersionNumber, Title, PublishStatus, or IsDeleted

ServiceNow: number, short_description, sys_mod_count, workflow_state, or active

 <p>The variable is replaced with the actual value for a piece of content when calling <a href="https://docs.aws.amazon.com/wisdom/latest/APIReference/API_GetContent.html">GetContent</a>. </p>

ResultData

resultData_relevanceScore :: Lens' ResultData (Maybe Double)

The relevance score of the results.

resultData_document :: Lens' ResultData Document

The document.

resultData_resultId :: Lens' ResultData Text

The identifier of the result data.

SearchExpression

searchExpression_filters :: Lens' SearchExpression [Filter]

The search expression filters.

ServerSideEncryptionConfiguration

serverSideEncryptionConfiguration_kmsKeyId :: Lens' ServerSideEncryptionConfiguration (Maybe Text)

The KMS key. For information about valid ID values, see Key identifiers (KeyId) in the AWS Key Management Service Developer Guide.

SessionData

sessionData_description :: Lens' SessionData (Maybe Text)

The description of the session.

sessionData_tags :: Lens' SessionData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

sessionData_name :: Lens' SessionData Text

The name of the session.

sessionData_sessionArn :: Lens' SessionData Text

The Amazon Resource Name (ARN) of the session.

sessionData_sessionId :: Lens' SessionData Text

The identifier of the session.

SessionSummary

sessionSummary_assistantArn :: Lens' SessionSummary Text

The Amazon Resource Name (ARN) of the Wisdom assistant

sessionSummary_assistantId :: Lens' SessionSummary Text

The identifier of the Wisdom assistant.

sessionSummary_sessionArn :: Lens' SessionSummary Text

The Amazon Resource Name (ARN) of the session.

sessionSummary_sessionId :: Lens' SessionSummary Text

The identifier of the session.

SourceConfiguration

sourceConfiguration_appIntegrations :: Lens' SourceConfiguration (Maybe AppIntegrationsConfiguration)

Configuration information for Amazon AppIntegrations to automatically ingest content.