Amazonka.Wisdom.ListContents

Description

Lists the content.

This operation returns paginated results.

Creating a Request

data ListContents

See: newListContents smart constructor.

Constructors

ListContents'

Fields

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per page.

knowledgeBaseId :: Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

newListContents

Arguments

:: Text
$sel:knowledgeBaseId:ListContents'

-> ListContents

Create a value of ListContents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContents', listContents_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListContents', listContents_maxResults - The maximum number of results to return per page.

$sel:knowledgeBaseId:ListContents', listContents_knowledgeBaseId - The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Request Lenses

listContents_nextToken :: Lens' ListContents (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listContents_maxResults :: Lens' ListContents (Maybe Natural)

The maximum number of results to return per page.

listContents_knowledgeBaseId :: Lens' ListContents Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Destructuring the Response

data ListContentsResponse

See: newListContentsResponse smart constructor.

Constructors

ListContentsResponse'

Fields

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

httpStatus :: Int

The response's http status code.

contentSummaries :: [ContentSummary]

Information about the content.

newListContentsResponse

Arguments

:: Int
$sel:httpStatus:ListContentsResponse'

-> ListContentsResponse

Create a value of ListContentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContents', listContentsResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:httpStatus:ListContentsResponse', listContentsResponse_httpStatus - The response's http status code.

$sel:contentSummaries:ListContentsResponse', listContentsResponse_contentSummaries - Information about the content.

Response Lenses

listContentsResponse_nextToken :: Lens' ListContentsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listContentsResponse_httpStatus :: Lens' ListContentsResponse Int

The response's http status code.

listContentsResponse_contentSummaries :: Lens' ListContentsResponse [ContentSummary]

Information about the content.