Amazonka.Wisdom.NotifyRecommendationsReceived

Description

Removes the specified recommendations from the specified assistant's queue of newly available recommendations. You can use this API in conjunction with GetRecommendations and a waitTimeSeconds input for long-polling behavior and avoiding duplicate recommendations.

Creating a Request

data NotifyRecommendationsReceived

See: newNotifyRecommendationsReceived smart constructor.

Constructors

NotifyRecommendationsReceived'

Fields

assistantId :: Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

recommendationIds :: [Text]

The identifiers of the recommendations.

sessionId :: Text

The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

newNotifyRecommendationsReceived

Arguments

:: Text
$sel:assistantId:NotifyRecommendationsReceived'

-> Text
$sel:sessionId:NotifyRecommendationsReceived'

-> NotifyRecommendationsReceived

Create a value of NotifyRecommendationsReceived with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assistantId:NotifyRecommendationsReceived', notifyRecommendationsReceived_assistantId - The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

$sel:recommendationIds:NotifyRecommendationsReceived', notifyRecommendationsReceived_recommendationIds - The identifiers of the recommendations.

$sel:sessionId:NotifyRecommendationsReceived', notifyRecommendationsReceived_sessionId - The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

Request Lenses

notifyRecommendationsReceived_assistantId :: Lens' NotifyRecommendationsReceived Text

The identifier of the Wisdom assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

notifyRecommendationsReceived_recommendationIds :: Lens' NotifyRecommendationsReceived [Text]

The identifiers of the recommendations.

notifyRecommendationsReceived_sessionId :: Lens' NotifyRecommendationsReceived Text

The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

Destructuring the Response

data NotifyRecommendationsReceivedResponse

See: newNotifyRecommendationsReceivedResponse smart constructor.

Constructors

NotifyRecommendationsReceivedResponse'

Fields

recommendationIds :: Maybe [Text]

The identifiers of the recommendations.

errors :: Maybe [NotifyRecommendationsReceivedError]

The identifiers of recommendations that are causing errors.

httpStatus :: Int

The response's http status code.

newNotifyRecommendationsReceivedResponse

Arguments

:: Int
$sel:httpStatus:NotifyRecommendationsReceivedResponse'

-> NotifyRecommendationsReceivedResponse

Create a value of NotifyRecommendationsReceivedResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommendationIds:NotifyRecommendationsReceived', notifyRecommendationsReceivedResponse_recommendationIds - The identifiers of the recommendations.

$sel:errors:NotifyRecommendationsReceivedResponse', notifyRecommendationsReceivedResponse_errors - The identifiers of recommendations that are causing errors.

$sel:httpStatus:NotifyRecommendationsReceivedResponse', notifyRecommendationsReceivedResponse_httpStatus - The response's http status code.

Response Lenses

notifyRecommendationsReceivedResponse_recommendationIds :: Lens' NotifyRecommendationsReceivedResponse (Maybe [Text])

The identifiers of the recommendations.

notifyRecommendationsReceivedResponse_errors :: Lens' NotifyRecommendationsReceivedResponse (Maybe [NotifyRecommendationsReceivedError])

The identifiers of recommendations that are causing errors.

notifyRecommendationsReceivedResponse_httpStatus :: Lens' NotifyRecommendationsReceivedResponse Int

The response's http status code.