Amazonka.Wisdom.Types

Description

Service Configuration

defaultService :: Service

API version 2020-10-19 of the Amazon Connect Wisdom Service SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_PreconditionFailedException :: AsError a => Getting (First ServiceError) a ServiceError

The provided revisionId does not match, indicating the content has been modified since it was last read.

_TooManyTagsException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Connect Wisdom throws this exception if you have too many tags in your tag set.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request could not be processed because of conflict in the current state of the resource. For example, if you're using a Create API (such as CreateAssistant) that accepts name, a conflicting resource (usually with the same name) is being created or mutated.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You've exceeded your service quota. To perform the requested action, remove some of the relevant resources, or use service quotas to request a service quota increase.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

AssistantStatus

newtype AssistantStatus

Constructors

AssistantStatus'

Fields

fromAssistantStatus :: Text

Bundled Patterns

pattern AssistantStatus_ACTIVE :: AssistantStatus

pattern AssistantStatus_CREATE_FAILED :: AssistantStatus

pattern AssistantStatus_CREATE_IN_PROGRESS :: AssistantStatus

pattern AssistantStatus_DELETED :: AssistantStatus

pattern AssistantStatus_DELETE_FAILED :: AssistantStatus

pattern AssistantStatus_DELETE_IN_PROGRESS :: AssistantStatus

AssistantType

newtype AssistantType

Constructors

AssistantType'

Fields

fromAssistantType :: Text

Bundled Patterns

pattern AssistantType_AGENT :: AssistantType

AssociationType

newtype AssociationType

Constructors

AssociationType'

Fields

fromAssociationType :: Text

Bundled Patterns

pattern AssociationType_KNOWLEDGE_BASE :: AssociationType

ContentStatus

newtype ContentStatus

Constructors

ContentStatus'

Fields

fromContentStatus :: Text

Bundled Patterns

pattern ContentStatus_ACTIVE :: ContentStatus

pattern ContentStatus_CREATE_FAILED :: ContentStatus

pattern ContentStatus_CREATE_IN_PROGRESS :: ContentStatus

pattern ContentStatus_DELETED :: ContentStatus

pattern ContentStatus_DELETE_FAILED :: ContentStatus

pattern ContentStatus_DELETE_IN_PROGRESS :: ContentStatus

pattern ContentStatus_UPDATE_FAILED :: ContentStatus

FilterField

newtype FilterField

Constructors

FilterField'

Fields

fromFilterField :: Text

Bundled Patterns

pattern FilterField_NAME :: FilterField

FilterOperator

newtype FilterOperator

Constructors

FilterOperator'

Fields

fromFilterOperator :: Text

Bundled Patterns

pattern FilterOperator_EQUALS :: FilterOperator

KnowledgeBaseStatus

newtype KnowledgeBaseStatus

Constructors

KnowledgeBaseStatus'

Fields

fromKnowledgeBaseStatus :: Text

Bundled Patterns

pattern KnowledgeBaseStatus_ACTIVE :: KnowledgeBaseStatus

pattern KnowledgeBaseStatus_CREATE_FAILED :: KnowledgeBaseStatus

pattern KnowledgeBaseStatus_CREATE_IN_PROGRESS :: KnowledgeBaseStatus

pattern KnowledgeBaseStatus_DELETED :: KnowledgeBaseStatus

pattern KnowledgeBaseStatus_DELETE_FAILED :: KnowledgeBaseStatus

pattern KnowledgeBaseStatus_DELETE_IN_PROGRESS :: KnowledgeBaseStatus

KnowledgeBaseType

newtype KnowledgeBaseType

Constructors

KnowledgeBaseType'

Fields

fromKnowledgeBaseType :: Text

Bundled Patterns

pattern KnowledgeBaseType_CUSTOM :: KnowledgeBaseType

pattern KnowledgeBaseType_EXTERNAL :: KnowledgeBaseType

RelevanceLevel

newtype RelevanceLevel

Constructors

RelevanceLevel'

Fields

fromRelevanceLevel :: Text

Bundled Patterns

pattern RelevanceLevel_HIGH :: RelevanceLevel

pattern RelevanceLevel_LOW :: RelevanceLevel

pattern RelevanceLevel_MEDIUM :: RelevanceLevel

AppIntegrationsConfiguration

data AppIntegrationsConfiguration

Configuration information for Amazon AppIntegrations to automatically ingest content.

See: newAppIntegrationsConfiguration smart constructor.

Constructors

AppIntegrationsConfiguration'

Fields

appIntegrationArn :: Text

The Amazon Resource Name (ARN) of the AppIntegrations DataIntegration to use for ingesting content.

objectFields :: NonEmpty Text

The fields from the source that are made available to your agents in Wisdom.

For Salesforce, you must include at least Id, ArticleNumber, VersionNumber, Title, PublishStatus, and IsDeleted.

For ServiceNow, you must include at least number, short_description, sys_mod_count, workflow_state, and active.

Make sure to include additional field(s); these are indexed and used to source recommendations.

newAppIntegrationsConfiguration

Arguments

:: Text
$sel:appIntegrationArn:AppIntegrationsConfiguration'

-> NonEmpty Text
$sel:objectFields:AppIntegrationsConfiguration'

-> AppIntegrationsConfiguration

Create a value of AppIntegrationsConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appIntegrationArn:AppIntegrationsConfiguration', appIntegrationsConfiguration_appIntegrationArn - The Amazon Resource Name (ARN) of the AppIntegrations DataIntegration to use for ingesting content.

$sel:objectFields:AppIntegrationsConfiguration', appIntegrationsConfiguration_objectFields - The fields from the source that are made available to your agents in Wisdom.

For Salesforce, you must include at least Id, ArticleNumber, VersionNumber, Title, PublishStatus, and IsDeleted.

For ServiceNow, you must include at least number, short_description, sys_mod_count, workflow_state, and active.

Make sure to include additional field(s); these are indexed and used to source recommendations.

appIntegrationsConfiguration_appIntegrationArn :: Lens' AppIntegrationsConfiguration Text

The Amazon Resource Name (ARN) of the AppIntegrations DataIntegration to use for ingesting content.

appIntegrationsConfiguration_objectFields :: Lens' AppIntegrationsConfiguration (NonEmpty Text)

The fields from the source that are made available to your agents in Wisdom.

For Salesforce, you must include at least Id, ArticleNumber, VersionNumber, Title, PublishStatus, and IsDeleted.

For ServiceNow, you must include at least number, short_description, sys_mod_count, workflow_state, and active.

Make sure to include additional field(s); these are indexed and used to source recommendations.

AssistantAssociationData

data AssistantAssociationData

Information about the assistant association.

See: newAssistantAssociationData smart constructor.

Constructors

AssistantAssociationData'

Fields

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

assistantArn :: Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantAssociationArn :: Text

The Amazon Resource Name (ARN) of the assistant association.

assistantAssociationId :: Text

The identifier of the assistant association.

assistantId :: Text

The identifier of the Wisdom assistant.

associationData :: AssistantAssociationOutputData

A union type that currently has a single argument, the knowledge base ID.

associationType :: AssociationType

The type of association.

newAssistantAssociationData

Arguments

:: Text
$sel:assistantArn:AssistantAssociationData'

-> Text
$sel:assistantAssociationArn:AssistantAssociationData'

-> Text
$sel:assistantAssociationId:AssistantAssociationData'

-> Text
$sel:assistantId:AssistantAssociationData'

-> AssistantAssociationOutputData
$sel:associationData:AssistantAssociationData'

-> AssociationType
$sel:associationType:AssistantAssociationData'

-> AssistantAssociationData

Create a value of AssistantAssociationData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:AssistantAssociationData', assistantAssociationData_tags - The tags used to organize, track, or control access for this resource.

$sel:assistantArn:AssistantAssociationData', assistantAssociationData_assistantArn - The Amazon Resource Name (ARN) of the Wisdom assistant

$sel:assistantAssociationArn:AssistantAssociationData', assistantAssociationData_assistantAssociationArn - The Amazon Resource Name (ARN) of the assistant association.

$sel:assistantAssociationId:AssistantAssociationData', assistantAssociationData_assistantAssociationId - The identifier of the assistant association.

$sel:assistantId:AssistantAssociationData', assistantAssociationData_assistantId - The identifier of the Wisdom assistant.

$sel:associationData:AssistantAssociationData', assistantAssociationData_associationData - A union type that currently has a single argument, the knowledge base ID.

$sel:associationType:AssistantAssociationData', assistantAssociationData_associationType - The type of association.

assistantAssociationData_tags :: Lens' AssistantAssociationData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

assistantAssociationData_assistantArn :: Lens' AssistantAssociationData Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantAssociationData_assistantAssociationArn :: Lens' AssistantAssociationData Text

The Amazon Resource Name (ARN) of the assistant association.

assistantAssociationData_assistantAssociationId :: Lens' AssistantAssociationData Text

The identifier of the assistant association.

assistantAssociationData_assistantId :: Lens' AssistantAssociationData Text

The identifier of the Wisdom assistant.

assistantAssociationData_associationData :: Lens' AssistantAssociationData AssistantAssociationOutputData

A union type that currently has a single argument, the knowledge base ID.

assistantAssociationData_associationType :: Lens' AssistantAssociationData AssociationType

The type of association.

AssistantAssociationInputData

data AssistantAssociationInputData

The data that is input into Wisdom as a result of the assistant association.

See: newAssistantAssociationInputData smart constructor.

Constructors

AssistantAssociationInputData'

Fields

knowledgeBaseId :: Maybe Text

The the identifier of the knowledge base.

newAssistantAssociationInputData :: AssistantAssociationInputData

Create a value of AssistantAssociationInputData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:knowledgeBaseId:AssistantAssociationInputData', assistantAssociationInputData_knowledgeBaseId - The the identifier of the knowledge base.

assistantAssociationInputData_knowledgeBaseId :: Lens' AssistantAssociationInputData (Maybe Text)

The the identifier of the knowledge base.

AssistantAssociationOutputData

data AssistantAssociationOutputData

The data that is output as a result of the assistant association.

See: newAssistantAssociationOutputData smart constructor.

Constructors

AssistantAssociationOutputData'

Fields

knowledgeBaseAssociation :: Maybe KnowledgeBaseAssociationData

The knowledge base where output data is sent.

newAssistantAssociationOutputData :: AssistantAssociationOutputData

Create a value of AssistantAssociationOutputData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:knowledgeBaseAssociation:AssistantAssociationOutputData', assistantAssociationOutputData_knowledgeBaseAssociation - The knowledge base where output data is sent.

assistantAssociationOutputData_knowledgeBaseAssociation :: Lens' AssistantAssociationOutputData (Maybe KnowledgeBaseAssociationData)

The knowledge base where output data is sent.

AssistantAssociationSummary

data AssistantAssociationSummary

Summary information about the assistant association.

See: newAssistantAssociationSummary smart constructor.

Constructors

AssistantAssociationSummary'

Fields

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

assistantArn :: Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantAssociationArn :: Text

The Amazon Resource Name (ARN) of the assistant association.

assistantAssociationId :: Text

The identifier of the assistant association.

assistantId :: Text

The identifier of the Wisdom assistant.

associationData :: AssistantAssociationOutputData

The association data.

associationType :: AssociationType

The type of association.

newAssistantAssociationSummary

Arguments

:: Text
$sel:assistantArn:AssistantAssociationSummary'

-> Text
$sel:assistantAssociationArn:AssistantAssociationSummary'

-> Text
$sel:assistantAssociationId:AssistantAssociationSummary'

-> Text
$sel:assistantId:AssistantAssociationSummary'

-> AssistantAssociationOutputData
$sel:associationData:AssistantAssociationSummary'

-> AssociationType
$sel:associationType:AssistantAssociationSummary'

-> AssistantAssociationSummary

Create a value of AssistantAssociationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:AssistantAssociationSummary', assistantAssociationSummary_tags - The tags used to organize, track, or control access for this resource.

$sel:assistantArn:AssistantAssociationSummary', assistantAssociationSummary_assistantArn - The Amazon Resource Name (ARN) of the Wisdom assistant

$sel:assistantAssociationArn:AssistantAssociationSummary', assistantAssociationSummary_assistantAssociationArn - The Amazon Resource Name (ARN) of the assistant association.

$sel:assistantAssociationId:AssistantAssociationSummary', assistantAssociationSummary_assistantAssociationId - The identifier of the assistant association.

$sel:assistantId:AssistantAssociationSummary', assistantAssociationSummary_assistantId - The identifier of the Wisdom assistant.

$sel:associationData:AssistantAssociationSummary', assistantAssociationSummary_associationData - The association data.

$sel:associationType:AssistantAssociationSummary', assistantAssociationSummary_associationType - The type of association.

assistantAssociationSummary_tags :: Lens' AssistantAssociationSummary (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

assistantAssociationSummary_assistantArn :: Lens' AssistantAssociationSummary Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantAssociationSummary_assistantAssociationArn :: Lens' AssistantAssociationSummary Text

The Amazon Resource Name (ARN) of the assistant association.

assistantAssociationSummary_assistantAssociationId :: Lens' AssistantAssociationSummary Text

The identifier of the assistant association.

assistantAssociationSummary_assistantId :: Lens' AssistantAssociationSummary Text

The identifier of the Wisdom assistant.

assistantAssociationSummary_associationData :: Lens' AssistantAssociationSummary AssistantAssociationOutputData

The association data.

assistantAssociationSummary_associationType :: Lens' AssistantAssociationSummary AssociationType

The type of association.

AssistantData

data AssistantData

The assistant data.

See: newAssistantData smart constructor.

Constructors

AssistantData'

Fields

description :: Maybe Text

The description.

serverSideEncryptionConfiguration :: Maybe ServerSideEncryptionConfiguration

The KMS key used for encryption.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

assistantArn :: Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantId :: Text

The identifier of the Wisdom assistant.

name :: Text

The name.

status :: AssistantStatus

The status of the assistant.

type' :: AssistantType

The type of assistant.

newAssistantData

Arguments

:: Text
$sel:assistantArn:AssistantData'

-> Text
$sel:assistantId:AssistantData'

-> Text
$sel:name:AssistantData'

-> AssistantStatus
$sel:status:AssistantData'

-> AssistantType
$sel:type':AssistantData'

-> AssistantData

Create a value of AssistantData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:AssistantData', assistantData_description - The description.

$sel:serverSideEncryptionConfiguration:AssistantData', assistantData_serverSideEncryptionConfiguration - The KMS key used for encryption.

$sel:tags:AssistantData', assistantData_tags - The tags used to organize, track, or control access for this resource.

$sel:assistantArn:AssistantData', assistantData_assistantArn - The Amazon Resource Name (ARN) of the Wisdom assistant

$sel:assistantId:AssistantData', assistantData_assistantId - The identifier of the Wisdom assistant.

$sel:name:AssistantData', assistantData_name - The name.

$sel:status:AssistantData', assistantData_status - The status of the assistant.

$sel:type':AssistantData', assistantData_type - The type of assistant.

assistantData_description :: Lens' AssistantData (Maybe Text)

The description.

assistantData_serverSideEncryptionConfiguration :: Lens' AssistantData (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

assistantData_tags :: Lens' AssistantData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

assistantData_assistantArn :: Lens' AssistantData Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantData_assistantId :: Lens' AssistantData Text

The identifier of the Wisdom assistant.

assistantData_name :: Lens' AssistantData Text

The name.

assistantData_status :: Lens' AssistantData AssistantStatus

The status of the assistant.

assistantData_type :: Lens' AssistantData AssistantType

The type of assistant.

AssistantSummary

data AssistantSummary

Summary information about the assistant.

See: newAssistantSummary smart constructor.

Constructors

AssistantSummary'

Fields

description :: Maybe Text

The description of the assistant.

serverSideEncryptionConfiguration :: Maybe ServerSideEncryptionConfiguration

The KMS key used for encryption.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

assistantArn :: Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantId :: Text

The identifier of the Wisdom assistant.

name :: Text

The name of the assistant.

status :: AssistantStatus

The status of the assistant.

type' :: AssistantType

The type of the assistant.

newAssistantSummary

Arguments

:: Text
$sel:assistantArn:AssistantSummary'

-> Text
$sel:assistantId:AssistantSummary'

-> Text
$sel:name:AssistantSummary'

-> AssistantStatus
$sel:status:AssistantSummary'

-> AssistantType
$sel:type':AssistantSummary'

-> AssistantSummary

Create a value of AssistantSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:AssistantSummary', assistantSummary_description - The description of the assistant.

$sel:serverSideEncryptionConfiguration:AssistantSummary', assistantSummary_serverSideEncryptionConfiguration - The KMS key used for encryption.

$sel:tags:AssistantSummary', assistantSummary_tags - The tags used to organize, track, or control access for this resource.

$sel:assistantArn:AssistantSummary', assistantSummary_assistantArn - The Amazon Resource Name (ARN) of the Wisdom assistant

$sel:assistantId:AssistantSummary', assistantSummary_assistantId - The identifier of the Wisdom assistant.

$sel:name:AssistantSummary', assistantSummary_name - The name of the assistant.

$sel:status:AssistantSummary', assistantSummary_status - The status of the assistant.

$sel:type':AssistantSummary', assistantSummary_type - The type of the assistant.

assistantSummary_description :: Lens' AssistantSummary (Maybe Text)

The description of the assistant.

assistantSummary_serverSideEncryptionConfiguration :: Lens' AssistantSummary (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

assistantSummary_tags :: Lens' AssistantSummary (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

assistantSummary_assistantArn :: Lens' AssistantSummary Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantSummary_assistantId :: Lens' AssistantSummary Text

The identifier of the Wisdom assistant.

assistantSummary_name :: Lens' AssistantSummary Text

The name of the assistant.

assistantSummary_status :: Lens' AssistantSummary AssistantStatus

The status of the assistant.

assistantSummary_type :: Lens' AssistantSummary AssistantType

The type of the assistant.

ContentData

data ContentData

Information about the content.

See: newContentData smart constructor.

Constructors

ContentData'

Fields

linkOutUri :: Maybe Text

The URI of the content.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

contentArn :: Text

The Amazon Resource Name (ARN) of the content.

contentId :: Text

The identifier of the content.

contentType :: Text

The media type of the content.

knowledgeBaseArn :: Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseId :: Text

The the identifier of the knowledge base.

metadata :: HashMap Text Text

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

name :: Text

The name of the content.

revisionId :: Text

The identifier of the content revision.

status :: ContentStatus

The status of the content.

title :: Text

The title of the content.

url :: Sensitive Text

The URL of the content.

urlExpiry :: POSIX

The expiration time of the URL as an epoch timestamp.

newContentData

Arguments

:: Text
$sel:contentArn:ContentData'

-> Text
$sel:contentId:ContentData'

-> Text
$sel:contentType:ContentData'

-> Text
$sel:knowledgeBaseArn:ContentData'

-> Text
$sel:knowledgeBaseId:ContentData'

-> Text
$sel:name:ContentData'

-> Text
$sel:revisionId:ContentData'

-> ContentStatus
$sel:status:ContentData'

-> Text
$sel:title:ContentData'

-> Text
$sel:url:ContentData'

-> UTCTime
$sel:urlExpiry:ContentData'

-> ContentData

Create a value of ContentData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:linkOutUri:ContentData', contentData_linkOutUri - The URI of the content.

$sel:tags:ContentData', contentData_tags - The tags used to organize, track, or control access for this resource.

$sel:contentArn:ContentData', contentData_contentArn - The Amazon Resource Name (ARN) of the content.

$sel:contentId:ContentData', contentData_contentId - The identifier of the content.

$sel:contentType:ContentData', contentData_contentType - The media type of the content.

$sel:knowledgeBaseArn:ContentData', contentData_knowledgeBaseArn - The Amazon Resource Name (ARN) of the knowledge base.

$sel:knowledgeBaseId:ContentData', contentData_knowledgeBaseId - The the identifier of the knowledge base.

$sel:metadata:ContentData', contentData_metadata - A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

$sel:name:ContentData', contentData_name - The name of the content.

$sel:revisionId:ContentData', contentData_revisionId - The identifier of the content revision.

$sel:status:ContentData', contentData_status - The status of the content.

$sel:title:ContentData', contentData_title - The title of the content.

$sel:url:ContentData', contentData_url - The URL of the content.

$sel:urlExpiry:ContentData', contentData_urlExpiry - The expiration time of the URL as an epoch timestamp.

contentData_linkOutUri :: Lens' ContentData (Maybe Text)

The URI of the content.

contentData_tags :: Lens' ContentData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

contentData_contentArn :: Lens' ContentData Text

The Amazon Resource Name (ARN) of the content.

contentData_contentId :: Lens' ContentData Text

The identifier of the content.

contentData_contentType :: Lens' ContentData Text

The media type of the content.

contentData_knowledgeBaseArn :: Lens' ContentData Text

The Amazon Resource Name (ARN) of the knowledge base.

contentData_knowledgeBaseId :: Lens' ContentData Text

The the identifier of the knowledge base.

contentData_metadata :: Lens' ContentData (HashMap Text Text)

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

contentData_name :: Lens' ContentData Text

The name of the content.

contentData_revisionId :: Lens' ContentData Text

The identifier of the content revision.

contentData_status :: Lens' ContentData ContentStatus

The status of the content.

contentData_title :: Lens' ContentData Text

The title of the content.

contentData_url :: Lens' ContentData Text

The URL of the content.

contentData_urlExpiry :: Lens' ContentData UTCTime

The expiration time of the URL as an epoch timestamp.

ContentReference

data ContentReference

Reference information about the content.

See: newContentReference smart constructor.

Constructors

ContentReference'

Fields

contentArn :: Maybe Text

The Amazon Resource Name (ARN) of the content.

contentId :: Maybe Text

The identifier of the content.

knowledgeBaseArn :: Maybe Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseId :: Maybe Text

The the identifier of the knowledge base.

newContentReference :: ContentReference

Create a value of ContentReference with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentArn:ContentReference', contentReference_contentArn - The Amazon Resource Name (ARN) of the content.

$sel:contentId:ContentReference', contentReference_contentId - The identifier of the content.

$sel:knowledgeBaseArn:ContentReference', contentReference_knowledgeBaseArn - The Amazon Resource Name (ARN) of the knowledge base.

$sel:knowledgeBaseId:ContentReference', contentReference_knowledgeBaseId - The the identifier of the knowledge base.

contentReference_contentArn :: Lens' ContentReference (Maybe Text)

The Amazon Resource Name (ARN) of the content.

contentReference_contentId :: Lens' ContentReference (Maybe Text)

The identifier of the content.

contentReference_knowledgeBaseArn :: Lens' ContentReference (Maybe Text)

The Amazon Resource Name (ARN) of the knowledge base.

contentReference_knowledgeBaseId :: Lens' ContentReference (Maybe Text)

The the identifier of the knowledge base.

ContentSummary

data ContentSummary

Summary information about the content.

See: newContentSummary smart constructor.

Constructors

ContentSummary'

Fields

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

contentArn :: Text

The Amazon Resource Name (ARN) of the content.

contentId :: Text

The identifier of the content.

contentType :: Text

The media type of the content.

knowledgeBaseArn :: Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseId :: Text

The the identifier of the knowledge base.

metadata :: HashMap Text Text

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

name :: Text

The name of the content.

revisionId :: Text

The identifier of the revision of the content.

status :: ContentStatus

The status of the content.

title :: Text

The title of the content.

newContentSummary

Arguments

:: Text
$sel:contentArn:ContentSummary'

-> Text
$sel:contentId:ContentSummary'

-> Text
$sel:contentType:ContentSummary'

-> Text
$sel:knowledgeBaseArn:ContentSummary'

-> Text
$sel:knowledgeBaseId:ContentSummary'

-> Text
$sel:name:ContentSummary'

-> Text
$sel:revisionId:ContentSummary'

-> ContentStatus
$sel:status:ContentSummary'

-> Text
$sel:title:ContentSummary'

-> ContentSummary

Create a value of ContentSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ContentSummary', contentSummary_tags - The tags used to organize, track, or control access for this resource.

$sel:contentArn:ContentSummary', contentSummary_contentArn - The Amazon Resource Name (ARN) of the content.

$sel:contentId:ContentSummary', contentSummary_contentId - The identifier of the content.

$sel:contentType:ContentSummary', contentSummary_contentType - The media type of the content.

$sel:knowledgeBaseArn:ContentSummary', contentSummary_knowledgeBaseArn - The Amazon Resource Name (ARN) of the knowledge base.

$sel:knowledgeBaseId:ContentSummary', contentSummary_knowledgeBaseId - The the identifier of the knowledge base.

$sel:metadata:ContentSummary', contentSummary_metadata - A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

$sel:name:ContentSummary', contentSummary_name - The name of the content.

$sel:revisionId:ContentSummary', contentSummary_revisionId - The identifier of the revision of the content.

$sel:status:ContentSummary', contentSummary_status - The status of the content.

$sel:title:ContentSummary', contentSummary_title - The title of the content.

contentSummary_tags :: Lens' ContentSummary (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

contentSummary_contentArn :: Lens' ContentSummary Text

The Amazon Resource Name (ARN) of the content.

contentSummary_contentId :: Lens' ContentSummary Text

The identifier of the content.

contentSummary_contentType :: Lens' ContentSummary Text

The media type of the content.

contentSummary_knowledgeBaseArn :: Lens' ContentSummary Text

The Amazon Resource Name (ARN) of the knowledge base.

contentSummary_knowledgeBaseId :: Lens' ContentSummary Text

The the identifier of the knowledge base.

contentSummary_metadata :: Lens' ContentSummary (HashMap Text Text)

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

contentSummary_name :: Lens' ContentSummary Text

The name of the content.

contentSummary_revisionId :: Lens' ContentSummary Text

The identifier of the revision of the content.

contentSummary_status :: Lens' ContentSummary ContentStatus

The status of the content.

contentSummary_title :: Lens' ContentSummary Text

The title of the content.

Document

data Document

The document.

See: newDocument smart constructor.

Constructors

Document'

Fields

excerpt :: Maybe DocumentText

The excerpt from the document.

title :: Maybe DocumentText

The title of the document.

contentReference :: ContentReference

A reference to the content resource.

newDocument

Arguments

:: ContentReference
$sel:contentReference:Document'

-> Document

Create a value of Document with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:excerpt:Document', document_excerpt - The excerpt from the document.

$sel:title:Document', document_title - The title of the document.

$sel:contentReference:Document', document_contentReference - A reference to the content resource.

document_excerpt :: Lens' Document (Maybe DocumentText)

The excerpt from the document.

document_title :: Lens' Document (Maybe DocumentText)

The title of the document.

document_contentReference :: Lens' Document ContentReference

A reference to the content resource.

DocumentText

data DocumentText

The text of the document.

See: newDocumentText smart constructor.

Constructors

DocumentText'

Fields

text :: Maybe (Sensitive Text)

Text in the document.

highlights :: Maybe [Highlight]

Highlights in the document text.

newDocumentText :: DocumentText

Create a value of DocumentText with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:text:DocumentText', documentText_text - Text in the document.

$sel:highlights:DocumentText', documentText_highlights - Highlights in the document text.

documentText_text :: Lens' DocumentText (Maybe Text)

Text in the document.

documentText_highlights :: Lens' DocumentText (Maybe [Highlight])

Highlights in the document text.

Filter

data Filter

A search filter.

See: newFilter smart constructor.

Constructors

Filter'

Fields

field :: FilterField

The field on which to filter.

operator :: FilterOperator

The operator to use for comparing the field’s value with the provided value.

value :: Text

The desired field value on which to filter.

newFilter

Arguments

:: FilterField
$sel:field:Filter'

-> FilterOperator
$sel:operator:Filter'

-> Text
$sel:value:Filter'

-> Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:field:Filter', filter_field - The field on which to filter.

$sel:operator:Filter', filter_operator - The operator to use for comparing the field’s value with the provided value.

$sel:value:Filter', filter_value - The desired field value on which to filter.

filter_field :: Lens' Filter FilterField

The field on which to filter.

filter_operator :: Lens' Filter FilterOperator

The operator to use for comparing the field’s value with the provided value.

filter_value :: Lens' Filter Text

The desired field value on which to filter.

Highlight

data Highlight

Offset specification to describe highlighting of document excerpts for rendering search results and recommendations.

See: newHighlight smart constructor.

Constructors

Highlight'

Fields

endOffsetExclusive :: Maybe Int

The offset for the end of the highlight.

beginOffsetInclusive :: Maybe Int

The offset for the start of the highlight.

newHighlight :: Highlight

Create a value of Highlight with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endOffsetExclusive:Highlight', highlight_endOffsetExclusive - The offset for the end of the highlight.

$sel:beginOffsetInclusive:Highlight', highlight_beginOffsetInclusive - The offset for the start of the highlight.

highlight_endOffsetExclusive :: Lens' Highlight (Maybe Int)

The offset for the end of the highlight.

highlight_beginOffsetInclusive :: Lens' Highlight (Maybe Int)

The offset for the start of the highlight.

KnowledgeBaseAssociationData

data KnowledgeBaseAssociationData

Association information about the knowledge base.

See: newKnowledgeBaseAssociationData smart constructor.

Constructors

KnowledgeBaseAssociationData'

Fields

knowledgeBaseArn :: Maybe Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseId :: Maybe Text

The the identifier of the knowledge base.

newKnowledgeBaseAssociationData :: KnowledgeBaseAssociationData

Create a value of KnowledgeBaseAssociationData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:knowledgeBaseArn:KnowledgeBaseAssociationData', knowledgeBaseAssociationData_knowledgeBaseArn - The Amazon Resource Name (ARN) of the knowledge base.

$sel:knowledgeBaseId:KnowledgeBaseAssociationData', knowledgeBaseAssociationData_knowledgeBaseId - The the identifier of the knowledge base.

knowledgeBaseAssociationData_knowledgeBaseArn :: Lens' KnowledgeBaseAssociationData (Maybe Text)

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseAssociationData_knowledgeBaseId :: Lens' KnowledgeBaseAssociationData (Maybe Text)

The the identifier of the knowledge base.

KnowledgeBaseData

data KnowledgeBaseData

Information about the knowledge base.

See: newKnowledgeBaseData smart constructor.

Constructors

KnowledgeBaseData'

Fields

renderingConfiguration :: Maybe RenderingConfiguration

Information about how to render the content.

sourceConfiguration :: Maybe SourceConfiguration

Source configuration information about the knowledge base.

lastContentModificationTime :: Maybe POSIX

An epoch timestamp indicating the most recent content modification inside the knowledge base. If no content exists in a knowledge base, this value is unset.

description :: Maybe Text

The description.

serverSideEncryptionConfiguration :: Maybe ServerSideEncryptionConfiguration

The KMS key used for encryption.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

knowledgeBaseArn :: Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseId :: Text

The the identifier of the knowledge base.

knowledgeBaseType :: KnowledgeBaseType

The type of knowledge base.

name :: Text

The name of the knowledge base.

status :: KnowledgeBaseStatus

The status of the knowledge base.

newKnowledgeBaseData

Arguments

:: Text
$sel:knowledgeBaseArn:KnowledgeBaseData'

-> Text
$sel:knowledgeBaseId:KnowledgeBaseData'

-> KnowledgeBaseType
$sel:knowledgeBaseType:KnowledgeBaseData'

-> Text
$sel:name:KnowledgeBaseData'

-> KnowledgeBaseStatus
$sel:status:KnowledgeBaseData'

-> KnowledgeBaseData

Create a value of KnowledgeBaseData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renderingConfiguration:KnowledgeBaseData', knowledgeBaseData_renderingConfiguration - Information about how to render the content.

$sel:sourceConfiguration:KnowledgeBaseData', knowledgeBaseData_sourceConfiguration - Source configuration information about the knowledge base.

$sel:lastContentModificationTime:KnowledgeBaseData', knowledgeBaseData_lastContentModificationTime - An epoch timestamp indicating the most recent content modification inside the knowledge base. If no content exists in a knowledge base, this value is unset.

$sel:description:KnowledgeBaseData', knowledgeBaseData_description - The description.

$sel:serverSideEncryptionConfiguration:KnowledgeBaseData', knowledgeBaseData_serverSideEncryptionConfiguration - The KMS key used for encryption.

$sel:tags:KnowledgeBaseData', knowledgeBaseData_tags - The tags used to organize, track, or control access for this resource.

$sel:knowledgeBaseArn:KnowledgeBaseData', knowledgeBaseData_knowledgeBaseArn - The Amazon Resource Name (ARN) of the knowledge base.

$sel:knowledgeBaseId:KnowledgeBaseData', knowledgeBaseData_knowledgeBaseId - The the identifier of the knowledge base.

$sel:knowledgeBaseType:KnowledgeBaseData', knowledgeBaseData_knowledgeBaseType - The type of knowledge base.

$sel:name:KnowledgeBaseData', knowledgeBaseData_name - The name of the knowledge base.

$sel:status:KnowledgeBaseData', knowledgeBaseData_status - The status of the knowledge base.

knowledgeBaseData_renderingConfiguration :: Lens' KnowledgeBaseData (Maybe RenderingConfiguration)

Information about how to render the content.

knowledgeBaseData_sourceConfiguration :: Lens' KnowledgeBaseData (Maybe SourceConfiguration)

Source configuration information about the knowledge base.

knowledgeBaseData_lastContentModificationTime :: Lens' KnowledgeBaseData (Maybe UTCTime)

An epoch timestamp indicating the most recent content modification inside the knowledge base. If no content exists in a knowledge base, this value is unset.

knowledgeBaseData_description :: Lens' KnowledgeBaseData (Maybe Text)

The description.

knowledgeBaseData_serverSideEncryptionConfiguration :: Lens' KnowledgeBaseData (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

knowledgeBaseData_tags :: Lens' KnowledgeBaseData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

knowledgeBaseData_knowledgeBaseArn :: Lens' KnowledgeBaseData Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseData_knowledgeBaseId :: Lens' KnowledgeBaseData Text

The the identifier of the knowledge base.

knowledgeBaseData_knowledgeBaseType :: Lens' KnowledgeBaseData KnowledgeBaseType

The type of knowledge base.

knowledgeBaseData_name :: Lens' KnowledgeBaseData Text

The name of the knowledge base.

knowledgeBaseData_status :: Lens' KnowledgeBaseData KnowledgeBaseStatus

The status of the knowledge base.

KnowledgeBaseSummary

data KnowledgeBaseSummary

Summary information about the knowledge base.

See: newKnowledgeBaseSummary smart constructor.

Constructors

KnowledgeBaseSummary'

Fields

renderingConfiguration :: Maybe RenderingConfiguration

Information about how to render the content.

sourceConfiguration :: Maybe SourceConfiguration

KEVIN

description :: Maybe Text

The description of the knowledge base.

serverSideEncryptionConfiguration :: Maybe ServerSideEncryptionConfiguration

The KMS key used for encryption.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

knowledgeBaseArn :: Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseId :: Text

The the identifier of the knowledge base.

knowledgeBaseType :: KnowledgeBaseType

The type of knowledge base.

name :: Text

The name of the knowledge base.

status :: KnowledgeBaseStatus

The status of the knowledge base summary.

newKnowledgeBaseSummary

Arguments

:: Text
$sel:knowledgeBaseArn:KnowledgeBaseSummary'

-> Text
$sel:knowledgeBaseId:KnowledgeBaseSummary'

-> KnowledgeBaseType
$sel:knowledgeBaseType:KnowledgeBaseSummary'

-> Text
$sel:name:KnowledgeBaseSummary'

-> KnowledgeBaseStatus
$sel:status:KnowledgeBaseSummary'

-> KnowledgeBaseSummary

Create a value of KnowledgeBaseSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renderingConfiguration:KnowledgeBaseSummary', knowledgeBaseSummary_renderingConfiguration - Information about how to render the content.

$sel:sourceConfiguration:KnowledgeBaseSummary', knowledgeBaseSummary_sourceConfiguration - [KEVIN]

$sel:description:KnowledgeBaseSummary', knowledgeBaseSummary_description - The description of the knowledge base.

$sel:serverSideEncryptionConfiguration:KnowledgeBaseSummary', knowledgeBaseSummary_serverSideEncryptionConfiguration - The KMS key used for encryption.

$sel:tags:KnowledgeBaseSummary', knowledgeBaseSummary_tags - The tags used to organize, track, or control access for this resource.

$sel:knowledgeBaseArn:KnowledgeBaseSummary', knowledgeBaseSummary_knowledgeBaseArn - The Amazon Resource Name (ARN) of the knowledge base.

$sel:knowledgeBaseId:KnowledgeBaseSummary', knowledgeBaseSummary_knowledgeBaseId - The the identifier of the knowledge base.

$sel:knowledgeBaseType:KnowledgeBaseSummary', knowledgeBaseSummary_knowledgeBaseType - The type of knowledge base.

$sel:name:KnowledgeBaseSummary', knowledgeBaseSummary_name - The name of the knowledge base.

$sel:status:KnowledgeBaseSummary', knowledgeBaseSummary_status - The status of the knowledge base summary.

knowledgeBaseSummary_renderingConfiguration :: Lens' KnowledgeBaseSummary (Maybe RenderingConfiguration)

Information about how to render the content.

knowledgeBaseSummary_sourceConfiguration :: Lens' KnowledgeBaseSummary (Maybe SourceConfiguration)

KEVIN

knowledgeBaseSummary_description :: Lens' KnowledgeBaseSummary (Maybe Text)

The description of the knowledge base.

knowledgeBaseSummary_serverSideEncryptionConfiguration :: Lens' KnowledgeBaseSummary (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

knowledgeBaseSummary_tags :: Lens' KnowledgeBaseSummary (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

knowledgeBaseSummary_knowledgeBaseArn :: Lens' KnowledgeBaseSummary Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseSummary_knowledgeBaseId :: Lens' KnowledgeBaseSummary Text

The the identifier of the knowledge base.

knowledgeBaseSummary_knowledgeBaseType :: Lens' KnowledgeBaseSummary KnowledgeBaseType

The type of knowledge base.

knowledgeBaseSummary_name :: Lens' KnowledgeBaseSummary Text

The name of the knowledge base.

knowledgeBaseSummary_status :: Lens' KnowledgeBaseSummary KnowledgeBaseStatus

The status of the knowledge base summary.

NotifyRecommendationsReceivedError

data NotifyRecommendationsReceivedError

An error occurred when creating a recommendation.

See: newNotifyRecommendationsReceivedError smart constructor.

Constructors

NotifyRecommendationsReceivedError'

Fields

recommendationId :: Maybe Text

The identifier of the recommendation that is in error.

message :: Maybe Text

A recommendation is causing an error.

newNotifyRecommendationsReceivedError :: NotifyRecommendationsReceivedError

Create a value of NotifyRecommendationsReceivedError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommendationId:NotifyRecommendationsReceivedError', notifyRecommendationsReceivedError_recommendationId - The identifier of the recommendation that is in error.

$sel:message:NotifyRecommendationsReceivedError', notifyRecommendationsReceivedError_message - A recommendation is causing an error.

notifyRecommendationsReceivedError_recommendationId :: Lens' NotifyRecommendationsReceivedError (Maybe Text)

The identifier of the recommendation that is in error.

notifyRecommendationsReceivedError_message :: Lens' NotifyRecommendationsReceivedError (Maybe Text)

A recommendation is causing an error.

RecommendationData

data RecommendationData

Information about the recommendation.

See: newRecommendationData smart constructor.

Constructors

RecommendationData'

Fields

relevanceLevel :: Maybe RelevanceLevel

The relevance level of the recommendation.

relevanceScore :: Maybe Double

The relevance score of the recommendation.

document :: Document

The recommended document.

recommendationId :: Text

The identifier of the recommendation.

newRecommendationData

Arguments

:: Document
$sel:document:RecommendationData'

-> Text
$sel:recommendationId:RecommendationData'

-> RecommendationData

Create a value of RecommendationData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relevanceLevel:RecommendationData', recommendationData_relevanceLevel - The relevance level of the recommendation.

$sel:relevanceScore:RecommendationData', recommendationData_relevanceScore - The relevance score of the recommendation.

$sel:document:RecommendationData', recommendationData_document - The recommended document.

$sel:recommendationId:RecommendationData', recommendationData_recommendationId - The identifier of the recommendation.

recommendationData_relevanceLevel :: Lens' RecommendationData (Maybe RelevanceLevel)

The relevance level of the recommendation.

recommendationData_relevanceScore :: Lens' RecommendationData (Maybe Double)

The relevance score of the recommendation.

recommendationData_document :: Lens' RecommendationData Document

The recommended document.

recommendationData_recommendationId :: Lens' RecommendationData Text

The identifier of the recommendation.

RenderingConfiguration

data RenderingConfiguration

Information about how to render the content.

See: newRenderingConfiguration smart constructor.

Constructors

RenderingConfiguration'

Fields

templateUri :: Maybe Text

A URI template containing exactly one variable in ${variableName} format. This can only be set for EXTERNAL knowledge bases. For Salesforce and ServiceNow, the variable must be one of the following:

Salesforce: Id, ArticleNumber, VersionNumber, Title, PublishStatus, or IsDeleted

ServiceNow: number, short_description, sys_mod_count, workflow_state, or active

 <p>The variable is replaced with the actual value for a piece of content when calling <a href="https://docs.aws.amazon.com/wisdom/latest/APIReference/API_GetContent.html">GetContent</a>. </p>

newRenderingConfiguration :: RenderingConfiguration

Create a value of RenderingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateUri:RenderingConfiguration', renderingConfiguration_templateUri - A URI template containing exactly one variable in ${variableName} format. This can only be set for EXTERNAL knowledge bases. For Salesforce and ServiceNow, the variable must be one of the following:

Salesforce: Id, ArticleNumber, VersionNumber, Title, PublishStatus, or IsDeleted

ServiceNow: number, short_description, sys_mod_count, workflow_state, or active

 <p>The variable is replaced with the actual value for a piece of content when calling <a href="https://docs.aws.amazon.com/wisdom/latest/APIReference/API_GetContent.html">GetContent</a>. </p>

renderingConfiguration_templateUri :: Lens' RenderingConfiguration (Maybe Text)

A URI template containing exactly one variable in ${variableName} format. This can only be set for EXTERNAL knowledge bases. For Salesforce and ServiceNow, the variable must be one of the following:

Salesforce: Id, ArticleNumber, VersionNumber, Title, PublishStatus, or IsDeleted

ServiceNow: number, short_description, sys_mod_count, workflow_state, or active

 <p>The variable is replaced with the actual value for a piece of content when calling <a href="https://docs.aws.amazon.com/wisdom/latest/APIReference/API_GetContent.html">GetContent</a>. </p>

ResultData

data ResultData

Information about the result.

See: newResultData smart constructor.

Constructors

ResultData'

Fields

relevanceScore :: Maybe Double

The relevance score of the results.

document :: Document

The document.

resultId :: Text

The identifier of the result data.

newResultData

Arguments

:: Document
$sel:document:ResultData'

-> Text
$sel:resultId:ResultData'

-> ResultData

Create a value of ResultData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relevanceScore:ResultData', resultData_relevanceScore - The relevance score of the results.

$sel:document:ResultData', resultData_document - The document.

$sel:resultId:ResultData', resultData_resultId - The identifier of the result data.

resultData_relevanceScore :: Lens' ResultData (Maybe Double)

The relevance score of the results.

resultData_document :: Lens' ResultData Document

The document.

resultData_resultId :: Lens' ResultData Text

The identifier of the result data.

SearchExpression

data SearchExpression

The search expression.

See: newSearchExpression smart constructor.

Constructors

SearchExpression'

Fields

filters :: [Filter]

The search expression filters.

newSearchExpression :: SearchExpression

Create a value of SearchExpression with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:SearchExpression', searchExpression_filters - The search expression filters.

searchExpression_filters :: Lens' SearchExpression [Filter]

The search expression filters.

ServerSideEncryptionConfiguration

data ServerSideEncryptionConfiguration

The KMS key used for encryption.

See: newServerSideEncryptionConfiguration smart constructor.

Constructors

ServerSideEncryptionConfiguration'

Fields

kmsKeyId :: Maybe Text

The KMS key. For information about valid ID values, see Key identifiers (KeyId) in the AWS Key Management Service Developer Guide.

newServerSideEncryptionConfiguration :: ServerSideEncryptionConfiguration

Create a value of ServerSideEncryptionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:ServerSideEncryptionConfiguration', serverSideEncryptionConfiguration_kmsKeyId - The KMS key. For information about valid ID values, see Key identifiers (KeyId) in the AWS Key Management Service Developer Guide.

serverSideEncryptionConfiguration_kmsKeyId :: Lens' ServerSideEncryptionConfiguration (Maybe Text)

The KMS key. For information about valid ID values, see Key identifiers (KeyId) in the AWS Key Management Service Developer Guide.

SessionData

data SessionData

Information about the session.

See: newSessionData smart constructor.

Constructors

SessionData'

Fields

description :: Maybe Text

The description of the session.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

name :: Text

The name of the session.

sessionArn :: Text

The Amazon Resource Name (ARN) of the session.

sessionId :: Text

The identifier of the session.

newSessionData

Arguments

:: Text
$sel:name:SessionData'

-> Text
$sel:sessionArn:SessionData'

-> Text
$sel:sessionId:SessionData'

-> SessionData

Create a value of SessionData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:SessionData', sessionData_description - The description of the session.

$sel:tags:SessionData', sessionData_tags - The tags used to organize, track, or control access for this resource.

$sel:name:SessionData', sessionData_name - The name of the session.

$sel:sessionArn:SessionData', sessionData_sessionArn - The Amazon Resource Name (ARN) of the session.

$sel:sessionId:SessionData', sessionData_sessionId - The identifier of the session.

sessionData_description :: Lens' SessionData (Maybe Text)

The description of the session.

sessionData_tags :: Lens' SessionData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

sessionData_name :: Lens' SessionData Text

The name of the session.

sessionData_sessionArn :: Lens' SessionData Text

The Amazon Resource Name (ARN) of the session.

sessionData_sessionId :: Lens' SessionData Text

The identifier of the session.

SessionSummary

data SessionSummary

Summary information about the session.

See: newSessionSummary smart constructor.

Constructors

SessionSummary'

Fields

assistantArn :: Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantId :: Text

The identifier of the Wisdom assistant.

sessionArn :: Text

The Amazon Resource Name (ARN) of the session.

sessionId :: Text

The identifier of the session.

newSessionSummary

Arguments

:: Text
$sel:assistantArn:SessionSummary'

-> Text
$sel:assistantId:SessionSummary'

-> Text
$sel:sessionArn:SessionSummary'

-> Text
$sel:sessionId:SessionSummary'

-> SessionSummary

Create a value of SessionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assistantArn:SessionSummary', sessionSummary_assistantArn - The Amazon Resource Name (ARN) of the Wisdom assistant

$sel:assistantId:SessionSummary', sessionSummary_assistantId - The identifier of the Wisdom assistant.

$sel:sessionArn:SessionSummary', sessionSummary_sessionArn - The Amazon Resource Name (ARN) of the session.

$sel:sessionId:SessionSummary', sessionSummary_sessionId - The identifier of the session.

sessionSummary_assistantArn :: Lens' SessionSummary Text

The Amazon Resource Name (ARN) of the Wisdom assistant

sessionSummary_assistantId :: Lens' SessionSummary Text

The identifier of the Wisdom assistant.

sessionSummary_sessionArn :: Lens' SessionSummary Text

The Amazon Resource Name (ARN) of the session.

sessionSummary_sessionId :: Lens' SessionSummary Text

The identifier of the session.

SourceConfiguration

data SourceConfiguration

Configuration information about the external data source.

See: newSourceConfiguration smart constructor.

Constructors

SourceConfiguration'

Fields

appIntegrations :: Maybe AppIntegrationsConfiguration

Configuration information for Amazon AppIntegrations to automatically ingest content.

newSourceConfiguration :: SourceConfiguration

Create a value of SourceConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appIntegrations:SourceConfiguration', sourceConfiguration_appIntegrations - Configuration information for Amazon AppIntegrations to automatically ingest content.

sourceConfiguration_appIntegrations :: Lens' SourceConfiguration (Maybe AppIntegrationsConfiguration)

Configuration information for Amazon AppIntegrations to automatically ingest content.