Amazonka.Wisdom.Types.AssistantData

Description

Documentation

data AssistantData

The assistant data.

See: newAssistantData smart constructor.

Constructors

AssistantData'

Fields

description :: Maybe Text

The description.

serverSideEncryptionConfiguration :: Maybe ServerSideEncryptionConfiguration

The KMS key used for encryption.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

assistantArn :: Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantId :: Text

The identifier of the Wisdom assistant.

name :: Text

The name.

status :: AssistantStatus

The status of the assistant.

type' :: AssistantType

The type of assistant.

newAssistantData

Arguments

:: Text
$sel:assistantArn:AssistantData'

-> Text
$sel:assistantId:AssistantData'

-> Text
$sel:name:AssistantData'

-> AssistantStatus
$sel:status:AssistantData'

-> AssistantType
$sel:type':AssistantData'

-> AssistantData

Create a value of AssistantData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:AssistantData', assistantData_description - The description.

$sel:serverSideEncryptionConfiguration:AssistantData', assistantData_serverSideEncryptionConfiguration - The KMS key used for encryption.

$sel:tags:AssistantData', assistantData_tags - The tags used to organize, track, or control access for this resource.

$sel:assistantArn:AssistantData', assistantData_assistantArn - The Amazon Resource Name (ARN) of the Wisdom assistant

$sel:assistantId:AssistantData', assistantData_assistantId - The identifier of the Wisdom assistant.

$sel:name:AssistantData', assistantData_name - The name.

$sel:status:AssistantData', assistantData_status - The status of the assistant.

$sel:type':AssistantData', assistantData_type - The type of assistant.

assistantData_description :: Lens' AssistantData (Maybe Text)

The description.

assistantData_serverSideEncryptionConfiguration :: Lens' AssistantData (Maybe ServerSideEncryptionConfiguration)

The KMS key used for encryption.

assistantData_tags :: Lens' AssistantData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

assistantData_assistantArn :: Lens' AssistantData Text

The Amazon Resource Name (ARN) of the Wisdom assistant

assistantData_assistantId :: Lens' AssistantData Text

The identifier of the Wisdom assistant.

assistantData_name :: Lens' AssistantData Text

The name.

assistantData_status :: Lens' AssistantData AssistantStatus

The status of the assistant.

assistantData_type :: Lens' AssistantData AssistantType

The type of assistant.