Amazonka.Wisdom.Types.ContentData

Description

Documentation

data ContentData

Information about the content.

See: newContentData smart constructor.

Constructors

ContentData'

Fields

linkOutUri :: Maybe Text

The URI of the content.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for this resource.

contentArn :: Text

The Amazon Resource Name (ARN) of the content.

contentId :: Text

The identifier of the content.

contentType :: Text

The media type of the content.

knowledgeBaseArn :: Text

The Amazon Resource Name (ARN) of the knowledge base.

knowledgeBaseId :: Text

The the identifier of the knowledge base.

metadata :: HashMap Text Text

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

name :: Text

The name of the content.

revisionId :: Text

The identifier of the content revision.

status :: ContentStatus

The status of the content.

title :: Text

The title of the content.

url :: Sensitive Text

The URL of the content.

urlExpiry :: POSIX

The expiration time of the URL as an epoch timestamp.

newContentData

Arguments

:: Text
$sel:contentArn:ContentData'

-> Text
$sel:contentId:ContentData'

-> Text
$sel:contentType:ContentData'

-> Text
$sel:knowledgeBaseArn:ContentData'

-> Text
$sel:knowledgeBaseId:ContentData'

-> Text
$sel:name:ContentData'

-> Text
$sel:revisionId:ContentData'

-> ContentStatus
$sel:status:ContentData'

-> Text
$sel:title:ContentData'

-> Text
$sel:url:ContentData'

-> UTCTime
$sel:urlExpiry:ContentData'

-> ContentData

Create a value of ContentData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:linkOutUri:ContentData', contentData_linkOutUri - The URI of the content.

$sel:tags:ContentData', contentData_tags - The tags used to organize, track, or control access for this resource.

$sel:contentArn:ContentData', contentData_contentArn - The Amazon Resource Name (ARN) of the content.

$sel:contentId:ContentData', contentData_contentId - The identifier of the content.

$sel:contentType:ContentData', contentData_contentType - The media type of the content.

$sel:knowledgeBaseArn:ContentData', contentData_knowledgeBaseArn - The Amazon Resource Name (ARN) of the knowledge base.

$sel:knowledgeBaseId:ContentData', contentData_knowledgeBaseId - The the identifier of the knowledge base.

$sel:metadata:ContentData', contentData_metadata - A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

$sel:name:ContentData', contentData_name - The name of the content.

$sel:revisionId:ContentData', contentData_revisionId - The identifier of the content revision.

$sel:status:ContentData', contentData_status - The status of the content.

$sel:title:ContentData', contentData_title - The title of the content.

$sel:url:ContentData', contentData_url - The URL of the content.

$sel:urlExpiry:ContentData', contentData_urlExpiry - The expiration time of the URL as an epoch timestamp.

contentData_linkOutUri :: Lens' ContentData (Maybe Text)

The URI of the content.

contentData_tags :: Lens' ContentData (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for this resource.

contentData_contentArn :: Lens' ContentData Text

The Amazon Resource Name (ARN) of the content.

contentData_contentId :: Lens' ContentData Text

The identifier of the content.

contentData_contentType :: Lens' ContentData Text

The media type of the content.

contentData_knowledgeBaseArn :: Lens' ContentData Text

The Amazon Resource Name (ARN) of the knowledge base.

contentData_knowledgeBaseId :: Lens' ContentData Text

The the identifier of the knowledge base.

contentData_metadata :: Lens' ContentData (HashMap Text Text)

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

contentData_name :: Lens' ContentData Text

The name of the content.

contentData_revisionId :: Lens' ContentData Text

The identifier of the content revision.

contentData_status :: Lens' ContentData ContentStatus

The status of the content.

contentData_title :: Lens' ContentData Text

The title of the content.

contentData_url :: Lens' ContentData Text

The URL of the content.

contentData_urlExpiry :: Lens' ContentData UTCTime

The expiration time of the URL as an epoch timestamp.