Amazonka.Wisdom.UpdateContent

Description

Updates information about the content.

Creating a Request

data UpdateContent

See: newUpdateContent smart constructor.

Constructors

UpdateContent'

Fields

overrideLinkOutUri :: Maybe Text

The URI for the article. If the knowledge base has a templateUri, setting this argument overrides it for this piece of content. To remove an existing overrideLinkOurUri, exclude this argument and set removeOverrideLinkOutUri to true.

removeOverrideLinkOutUri :: Maybe Bool

Unset the existing overrideLinkOutUri if it exists.

metadata :: Maybe (HashMap Text Text)

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

title :: Maybe Text

The title of the content.

revisionId :: Maybe Text

The revisionId of the content resource to update, taken from an earlier call to GetContent, GetContentSummary, SearchContent, or ListContents. If included, this argument acts as an optimistic lock to ensure content was not modified since it was last read. If it has been modified, this API throws a PreconditionFailedException.

uploadId :: Maybe Text

A pointer to the uploaded asset. This value is returned by StartContentUpload.

contentId :: Text

The identifier of the content. Can be either the ID or the ARN. URLs cannot contain the ARN.

knowledgeBaseId :: Text

The the identifier of the knowledge base. Can be either the ID or the ARN

newUpdateContent

Arguments

:: Text
$sel:contentId:UpdateContent'

-> Text
$sel:knowledgeBaseId:UpdateContent'

-> UpdateContent

Create a value of UpdateContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:overrideLinkOutUri:UpdateContent', updateContent_overrideLinkOutUri - The URI for the article. If the knowledge base has a templateUri, setting this argument overrides it for this piece of content. To remove an existing overrideLinkOurUri, exclude this argument and set removeOverrideLinkOutUri to true.

$sel:removeOverrideLinkOutUri:UpdateContent', updateContent_removeOverrideLinkOutUri - Unset the existing overrideLinkOutUri if it exists.

$sel:metadata:UpdateContent', updateContent_metadata - A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

$sel:title:UpdateContent', updateContent_title - The title of the content.

$sel:revisionId:UpdateContent', updateContent_revisionId - The revisionId of the content resource to update, taken from an earlier call to GetContent, GetContentSummary, SearchContent, or ListContents. If included, this argument acts as an optimistic lock to ensure content was not modified since it was last read. If it has been modified, this API throws a PreconditionFailedException.

$sel:uploadId:UpdateContent', updateContent_uploadId - A pointer to the uploaded asset. This value is returned by StartContentUpload.

$sel:contentId:UpdateContent', updateContent_contentId - The identifier of the content. Can be either the ID or the ARN. URLs cannot contain the ARN.

$sel:knowledgeBaseId:UpdateContent', updateContent_knowledgeBaseId - The the identifier of the knowledge base. Can be either the ID or the ARN

Request Lenses

updateContent_overrideLinkOutUri :: Lens' UpdateContent (Maybe Text)

The URI for the article. If the knowledge base has a templateUri, setting this argument overrides it for this piece of content. To remove an existing overrideLinkOurUri, exclude this argument and set removeOverrideLinkOutUri to true.

updateContent_removeOverrideLinkOutUri :: Lens' UpdateContent (Maybe Bool)

Unset the existing overrideLinkOutUri if it exists.

updateContent_metadata :: Lens' UpdateContent (Maybe (HashMap Text Text))

A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Wisdom, you can store an external version identifier as metadata to utilize for determining drift.

updateContent_title :: Lens' UpdateContent (Maybe Text)

The title of the content.

updateContent_revisionId :: Lens' UpdateContent (Maybe Text)

The revisionId of the content resource to update, taken from an earlier call to GetContent, GetContentSummary, SearchContent, or ListContents. If included, this argument acts as an optimistic lock to ensure content was not modified since it was last read. If it has been modified, this API throws a PreconditionFailedException.

updateContent_uploadId :: Lens' UpdateContent (Maybe Text)

A pointer to the uploaded asset. This value is returned by StartContentUpload.

updateContent_contentId :: Lens' UpdateContent Text

The identifier of the content. Can be either the ID or the ARN. URLs cannot contain the ARN.

updateContent_knowledgeBaseId :: Lens' UpdateContent Text

The the identifier of the knowledge base. Can be either the ID or the ARN

Destructuring the Response

data UpdateContentResponse

See: newUpdateContentResponse smart constructor.

Constructors

UpdateContentResponse'

Fields

content :: Maybe ContentData

The content.

httpStatus :: Int

The response's http status code.

newUpdateContentResponse

Arguments

:: Int
$sel:httpStatus:UpdateContentResponse'

-> UpdateContentResponse

Create a value of UpdateContentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:content:UpdateContentResponse', updateContentResponse_content - The content.

$sel:httpStatus:UpdateContentResponse', updateContentResponse_httpStatus - The response's http status code.

Response Lenses

updateContentResponse_content :: Lens' UpdateContentResponse (Maybe ContentData)

The content.

updateContentResponse_httpStatus :: Lens' UpdateContentResponse Int

The response's http status code.