Amazonka.Wisdom.UpdateKnowledgeBaseTemplateUri

Description

Updates the template URI of a knowledge base. This is only supported for knowledge bases of type EXTERNAL. Include a single variable in ${variable} format; this interpolated by Wisdom using ingested content. For example, if you ingest a Salesforce article, it has an Id value, and you can set the template URI to https://myInstanceName.lightning.force.com/lightning/r/Knowledge__kav/*${Id}*/view.

Creating a Request

data UpdateKnowledgeBaseTemplateUri

See: newUpdateKnowledgeBaseTemplateUri smart constructor.

Constructors

UpdateKnowledgeBaseTemplateUri'

Fields

knowledgeBaseId :: Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

templateUri :: Text

The template URI to update.

newUpdateKnowledgeBaseTemplateUri

Arguments

:: Text
$sel:knowledgeBaseId:UpdateKnowledgeBaseTemplateUri'

-> Text
$sel:templateUri:UpdateKnowledgeBaseTemplateUri'

-> UpdateKnowledgeBaseTemplateUri

Create a value of UpdateKnowledgeBaseTemplateUri with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:knowledgeBaseId:UpdateKnowledgeBaseTemplateUri', updateKnowledgeBaseTemplateUri_knowledgeBaseId - The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

$sel:templateUri:UpdateKnowledgeBaseTemplateUri', updateKnowledgeBaseTemplateUri_templateUri - The template URI to update.

Request Lenses

updateKnowledgeBaseTemplateUri_knowledgeBaseId :: Lens' UpdateKnowledgeBaseTemplateUri Text

The the identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

updateKnowledgeBaseTemplateUri_templateUri :: Lens' UpdateKnowledgeBaseTemplateUri Text

The template URI to update.

Destructuring the Response

data UpdateKnowledgeBaseTemplateUriResponse

See: newUpdateKnowledgeBaseTemplateUriResponse smart constructor.

Constructors

UpdateKnowledgeBaseTemplateUriResponse'

Fields

knowledgeBase :: Maybe KnowledgeBaseData

The knowledge base to update.

httpStatus :: Int

The response's http status code.

newUpdateKnowledgeBaseTemplateUriResponse

Arguments

:: Int
$sel:httpStatus:UpdateKnowledgeBaseTemplateUriResponse'

-> UpdateKnowledgeBaseTemplateUriResponse

Create a value of UpdateKnowledgeBaseTemplateUriResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:knowledgeBase:UpdateKnowledgeBaseTemplateUriResponse', updateKnowledgeBaseTemplateUriResponse_knowledgeBase - The knowledge base to update.

$sel:httpStatus:UpdateKnowledgeBaseTemplateUriResponse', updateKnowledgeBaseTemplateUriResponse_httpStatus - The response's http status code.

Response Lenses

updateKnowledgeBaseTemplateUriResponse_knowledgeBase :: Lens' UpdateKnowledgeBaseTemplateUriResponse (Maybe KnowledgeBaseData)

The knowledge base to update.

updateKnowledgeBaseTemplateUriResponse_httpStatus :: Lens' UpdateKnowledgeBaseTemplateUriResponse Int

The response's http status code.