Amazonka.WorkDocs.GetDocument

Description

Retrieves details of a document.

Creating a Request

data GetDocument

See: newGetDocument smart constructor.

Constructors

GetDocument'

Fields

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

includeCustomMetadata :: Maybe Bool

Set this to TRUE to include custom metadata in the response.

documentId :: Text

The ID of the document.

newGetDocument

Arguments

:: Text
$sel:documentId:GetDocument'

-> GetDocument

Create a value of GetDocument with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authenticationToken:GetDocument', getDocument_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:includeCustomMetadata:GetDocument', getDocument_includeCustomMetadata - Set this to TRUE to include custom metadata in the response.

$sel:documentId:GetDocument', getDocument_documentId - The ID of the document.

Request Lenses

getDocument_authenticationToken :: Lens' GetDocument (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

getDocument_includeCustomMetadata :: Lens' GetDocument (Maybe Bool)

Set this to TRUE to include custom metadata in the response.

getDocument_documentId :: Lens' GetDocument Text

The ID of the document.

Destructuring the Response

data GetDocumentResponse

See: newGetDocumentResponse smart constructor.

Constructors

GetDocumentResponse'

Fields

customMetadata :: Maybe (HashMap Text Text)

The custom metadata on the document.

metadata :: Maybe DocumentMetadata

The metadata details of the document.

httpStatus :: Int

The response's http status code.

newGetDocumentResponse

Arguments

:: Int
$sel:httpStatus:GetDocumentResponse'

-> GetDocumentResponse

Create a value of GetDocumentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customMetadata:GetDocumentResponse', getDocumentResponse_customMetadata - The custom metadata on the document.

$sel:metadata:GetDocumentResponse', getDocumentResponse_metadata - The metadata details of the document.

$sel:httpStatus:GetDocumentResponse', getDocumentResponse_httpStatus - The response's http status code.

Response Lenses

getDocumentResponse_customMetadata :: Lens' GetDocumentResponse (Maybe (HashMap Text Text))

The custom metadata on the document.

getDocumentResponse_metadata :: Lens' GetDocumentResponse (Maybe DocumentMetadata)

The metadata details of the document.

getDocumentResponse_httpStatus :: Lens' GetDocumentResponse Int

The response's http status code.