Amazonka.WorkDocs.GetDocumentVersion

Description

Retrieves version metadata for the specified document.

Creating a Request

data GetDocumentVersion

See: newGetDocumentVersion smart constructor.

Constructors

GetDocumentVersion'

Fields

authenticationToken :: Maybe (Sensitive Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

includeCustomMetadata :: Maybe Bool

Set this to TRUE to include custom metadata in the response.

fields :: Maybe Text

A comma-separated list of values. Specify "SOURCE" to include a URL for the source document.

documentId :: Text

The ID of the document.

versionId :: Text

The version ID of the document.

newGetDocumentVersion

Arguments

:: Text
$sel:documentId:GetDocumentVersion'

-> Text
$sel:versionId:GetDocumentVersion'

-> GetDocumentVersion

Create a value of GetDocumentVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authenticationToken:GetDocumentVersion', getDocumentVersion_authenticationToken - Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

$sel:includeCustomMetadata:GetDocumentVersion', getDocumentVersion_includeCustomMetadata - Set this to TRUE to include custom metadata in the response.

$sel:fields:GetDocumentVersion', getDocumentVersion_fields - A comma-separated list of values. Specify "SOURCE" to include a URL for the source document.

$sel:documentId:GetDocumentVersion', getDocumentVersion_documentId - The ID of the document.

$sel:versionId:GetDocumentVersion', getDocumentVersion_versionId - The version ID of the document.

Request Lenses

getDocumentVersion_authenticationToken :: Lens' GetDocumentVersion (Maybe Text)

Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

getDocumentVersion_includeCustomMetadata :: Lens' GetDocumentVersion (Maybe Bool)

Set this to TRUE to include custom metadata in the response.

getDocumentVersion_fields :: Lens' GetDocumentVersion (Maybe Text)

A comma-separated list of values. Specify "SOURCE" to include a URL for the source document.

getDocumentVersion_documentId :: Lens' GetDocumentVersion Text

The ID of the document.

getDocumentVersion_versionId :: Lens' GetDocumentVersion Text

The version ID of the document.

Destructuring the Response

data GetDocumentVersionResponse

See: newGetDocumentVersionResponse smart constructor.

Constructors

GetDocumentVersionResponse'

Fields

customMetadata :: Maybe (HashMap Text Text)

The custom metadata on the document version.

metadata :: Maybe DocumentVersionMetadata

The version metadata.

httpStatus :: Int

The response's http status code.

newGetDocumentVersionResponse

Arguments

:: Int
$sel:httpStatus:GetDocumentVersionResponse'

-> GetDocumentVersionResponse

Create a value of GetDocumentVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customMetadata:GetDocumentVersionResponse', getDocumentVersionResponse_customMetadata - The custom metadata on the document version.

$sel:metadata:GetDocumentVersionResponse', getDocumentVersionResponse_metadata - The version metadata.

$sel:httpStatus:GetDocumentVersionResponse', getDocumentVersionResponse_httpStatus - The response's http status code.

Response Lenses

getDocumentVersionResponse_customMetadata :: Lens' GetDocumentVersionResponse (Maybe (HashMap Text Text))

The custom metadata on the document version.

getDocumentVersionResponse_metadata :: Lens' GetDocumentVersionResponse (Maybe DocumentVersionMetadata)

The version metadata.

getDocumentVersionResponse_httpStatus :: Lens' GetDocumentVersionResponse Int

The response's http status code.