Amazonka.WorkDocs.Types.Comment

Description

Documentation

data Comment

Describes a comment.

See: newComment smart constructor.

Constructors

Comment'

Fields

status :: Maybe CommentStatusType

The status of the comment.

text :: Maybe (Sensitive Text)

The text of the comment.

visibility :: Maybe CommentVisibilityType

The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

threadId :: Maybe Text

The ID of the root comment in the thread.

contributor :: Maybe User

The details of the user who made the comment.

createdTimestamp :: Maybe POSIX

The time that the comment was created.

recipientId :: Maybe Text

If the comment is a reply to another user's comment, this field contains the user ID of the user being replied to.

parentId :: Maybe Text

The ID of the parent comment.

commentId :: Text

The ID of the comment.

newComment

Arguments

:: Text
$sel:commentId:Comment'

-> Comment

Create a value of Comment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Comment', comment_status - The status of the comment.

$sel:text:Comment', comment_text - The text of the comment.

$sel:visibility:Comment', comment_visibility - The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

$sel:threadId:Comment', comment_threadId - The ID of the root comment in the thread.

$sel:contributor:Comment', comment_contributor - The details of the user who made the comment.

$sel:createdTimestamp:Comment', comment_createdTimestamp - The time that the comment was created.

$sel:recipientId:Comment', comment_recipientId - If the comment is a reply to another user's comment, this field contains the user ID of the user being replied to.

$sel:parentId:Comment', comment_parentId - The ID of the parent comment.

$sel:commentId:Comment', comment_commentId - The ID of the comment.

comment_status :: Lens' Comment (Maybe CommentStatusType)

The status of the comment.

comment_text :: Lens' Comment (Maybe Text)

The text of the comment.

comment_visibility :: Lens' Comment (Maybe CommentVisibilityType)

The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

comment_threadId :: Lens' Comment (Maybe Text)

The ID of the root comment in the thread.

comment_contributor :: Lens' Comment (Maybe User)

The details of the user who made the comment.

comment_createdTimestamp :: Lens' Comment (Maybe UTCTime)

The time that the comment was created.

comment_recipientId :: Lens' Comment (Maybe Text)

If the comment is a reply to another user's comment, this field contains the user ID of the user being replied to.

comment_parentId :: Lens' Comment (Maybe Text)

The ID of the parent comment.

comment_commentId :: Lens' Comment Text

The ID of the comment.