Amazonka.WorkLink.Types.FleetSummary

Description

Documentation

data FleetSummary

The summary of the fleet.

See: newFleetSummary smart constructor.

Constructors

FleetSummary'

Fields

lastUpdatedTime :: Maybe POSIX

The time when the fleet was last updated.

fleetStatus :: Maybe FleetStatus

The status of the fleet.

companyCode :: Maybe Text

The identifier used by users to sign into the Amazon WorkLink app.

createdTime :: Maybe POSIX

The time when the fleet was created.

fleetArn :: Maybe Text

The Amazon Resource Name (ARN) of the fleet.

displayName :: Maybe Text

The name of the fleet to display.

fleetName :: Maybe Text

The name of the fleet.

tags :: Maybe (HashMap Text Text)

The tags attached to the resource. A tag is a key-value pair.

newFleetSummary :: FleetSummary

Create a value of FleetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:FleetSummary', fleetSummary_lastUpdatedTime - The time when the fleet was last updated.

$sel:fleetStatus:FleetSummary', fleetSummary_fleetStatus - The status of the fleet.

$sel:companyCode:FleetSummary', fleetSummary_companyCode - The identifier used by users to sign into the Amazon WorkLink app.

$sel:createdTime:FleetSummary', fleetSummary_createdTime - The time when the fleet was created.

$sel:fleetArn:FleetSummary', fleetSummary_fleetArn - The Amazon Resource Name (ARN) of the fleet.

$sel:displayName:FleetSummary', fleetSummary_displayName - The name of the fleet to display.

$sel:fleetName:FleetSummary', fleetSummary_fleetName - The name of the fleet.

$sel:tags:FleetSummary', fleetSummary_tags - The tags attached to the resource. A tag is a key-value pair.

fleetSummary_lastUpdatedTime :: Lens' FleetSummary (Maybe UTCTime)

The time when the fleet was last updated.

fleetSummary_fleetStatus :: Lens' FleetSummary (Maybe FleetStatus)

The status of the fleet.

fleetSummary_companyCode :: Lens' FleetSummary (Maybe Text)

The identifier used by users to sign into the Amazon WorkLink app.

fleetSummary_createdTime :: Lens' FleetSummary (Maybe UTCTime)

The time when the fleet was created.

fleetSummary_fleetArn :: Lens' FleetSummary (Maybe Text)

The Amazon Resource Name (ARN) of the fleet.

fleetSummary_displayName :: Lens' FleetSummary (Maybe Text)

The name of the fleet to display.

fleetSummary_fleetName :: Lens' FleetSummary (Maybe Text)

The name of the fleet.

fleetSummary_tags :: Lens' FleetSummary (Maybe (HashMap Text Text))

The tags attached to the resource. A tag is a key-value pair.