Amazonka.WorkMail.DeregisterFromWorkMail

Description

Mark a user, group, or resource as no longer used in Amazon WorkMail. This action disassociates the mailbox and schedules it for clean-up. WorkMail keeps mailboxes for 30 days before they are permanently removed. The functionality in the console is Disable.

Creating a Request

data DeregisterFromWorkMail

See: newDeregisterFromWorkMail smart constructor.

Constructors

DeregisterFromWorkMail'

Fields

organizationId :: Text

The identifier for the organization under which the Amazon WorkMail entity exists.

entityId :: Text

The identifier for the member (user or group) to be updated.

newDeregisterFromWorkMail

Arguments

:: Text
$sel:organizationId:DeregisterFromWorkMail'

-> Text
$sel:entityId:DeregisterFromWorkMail'

-> DeregisterFromWorkMail

Create a value of DeregisterFromWorkMail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationId:DeregisterFromWorkMail', deregisterFromWorkMail_organizationId - The identifier for the organization under which the Amazon WorkMail entity exists.

$sel:entityId:DeregisterFromWorkMail', deregisterFromWorkMail_entityId - The identifier for the member (user or group) to be updated.

Request Lenses

deregisterFromWorkMail_organizationId :: Lens' DeregisterFromWorkMail Text

The identifier for the organization under which the Amazon WorkMail entity exists.

deregisterFromWorkMail_entityId :: Lens' DeregisterFromWorkMail Text

The identifier for the member (user or group) to be updated.

Destructuring the Response

data DeregisterFromWorkMailResponse

See: newDeregisterFromWorkMailResponse smart constructor.

Constructors

DeregisterFromWorkMailResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeregisterFromWorkMailResponse

Arguments

:: Int
$sel:httpStatus:DeregisterFromWorkMailResponse'

-> DeregisterFromWorkMailResponse

Create a value of DeregisterFromWorkMailResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeregisterFromWorkMailResponse', deregisterFromWorkMailResponse_httpStatus - The response's http status code.

Response Lenses

deregisterFromWorkMailResponse_httpStatus :: Lens' DeregisterFromWorkMailResponse Int

The response's http status code.