Amazonka.WorkMail.DescribeMailboxExportJob

Description

Describes the current status of a mailbox export job.

Creating a Request

data DescribeMailboxExportJob

See: newDescribeMailboxExportJob smart constructor.

Constructors

DescribeMailboxExportJob'

Fields

jobId :: Text

The mailbox export job ID.

organizationId :: Text

The organization ID.

newDescribeMailboxExportJob

Arguments

:: Text
$sel:jobId:DescribeMailboxExportJob'

-> Text
$sel:organizationId:DescribeMailboxExportJob'

-> DescribeMailboxExportJob

Create a value of DescribeMailboxExportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:DescribeMailboxExportJob', describeMailboxExportJob_jobId - The mailbox export job ID.

$sel:organizationId:DescribeMailboxExportJob', describeMailboxExportJob_organizationId - The organization ID.

Request Lenses

describeMailboxExportJob_jobId :: Lens' DescribeMailboxExportJob Text

The mailbox export job ID.

describeMailboxExportJob_organizationId :: Lens' DescribeMailboxExportJob Text

The organization ID.

Destructuring the Response

data DescribeMailboxExportJobResponse

See: newDescribeMailboxExportJobResponse smart constructor.

Constructors

DescribeMailboxExportJobResponse'

Fields

state :: Maybe MailboxExportJobState

The state of the mailbox export job.

kmsKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of the symmetric AWS Key Management Service (AWS KMS) key that encrypts the exported mailbox content.

startTime :: Maybe POSIX

The mailbox export job start timestamp.

estimatedProgress :: Maybe Natural

The estimated progress of the mailbox export job, in percentage points.

endTime :: Maybe POSIX

The mailbox export job end timestamp.

s3Path :: Maybe Text

The path to the S3 bucket and file that the mailbox export job is exporting to.

s3Prefix :: Maybe Text

The S3 bucket prefix.

entityId :: Maybe Text

The identifier of the user or resource associated with the mailbox.

description :: Maybe Text

The mailbox export job description.

errorInfo :: Maybe Text

Error information for failed mailbox export jobs.

s3BucketName :: Maybe Text

The name of the S3 bucket.

roleArn :: Maybe Text

The ARN of the AWS Identity and Access Management (IAM) role that grants write permission to the Amazon Simple Storage Service (Amazon S3) bucket.

httpStatus :: Int

The response's http status code.

newDescribeMailboxExportJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeMailboxExportJobResponse'

-> DescribeMailboxExportJobResponse

Create a value of DescribeMailboxExportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_state - The state of the mailbox export job.

$sel:kmsKeyArn:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_kmsKeyArn - The Amazon Resource Name (ARN) of the symmetric AWS Key Management Service (AWS KMS) key that encrypts the exported mailbox content.

$sel:startTime:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_startTime - The mailbox export job start timestamp.

$sel:estimatedProgress:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_estimatedProgress - The estimated progress of the mailbox export job, in percentage points.

$sel:endTime:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_endTime - The mailbox export job end timestamp.

$sel:s3Path:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_s3Path - The path to the S3 bucket and file that the mailbox export job is exporting to.

$sel:s3Prefix:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_s3Prefix - The S3 bucket prefix.

$sel:entityId:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_entityId - The identifier of the user or resource associated with the mailbox.

$sel:description:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_description - The mailbox export job description.

$sel:errorInfo:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_errorInfo - Error information for failed mailbox export jobs.

$sel:s3BucketName:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_s3BucketName - The name of the S3 bucket.

$sel:roleArn:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_roleArn - The ARN of the AWS Identity and Access Management (IAM) role that grants write permission to the Amazon Simple Storage Service (Amazon S3) bucket.

$sel:httpStatus:DescribeMailboxExportJobResponse', describeMailboxExportJobResponse_httpStatus - The response's http status code.

Response Lenses

describeMailboxExportJobResponse_state :: Lens' DescribeMailboxExportJobResponse (Maybe MailboxExportJobState)

The state of the mailbox export job.

describeMailboxExportJobResponse_kmsKeyArn :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the symmetric AWS Key Management Service (AWS KMS) key that encrypts the exported mailbox content.

describeMailboxExportJobResponse_startTime :: Lens' DescribeMailboxExportJobResponse (Maybe UTCTime)

The mailbox export job start timestamp.

describeMailboxExportJobResponse_estimatedProgress :: Lens' DescribeMailboxExportJobResponse (Maybe Natural)

The estimated progress of the mailbox export job, in percentage points.

describeMailboxExportJobResponse_endTime :: Lens' DescribeMailboxExportJobResponse (Maybe UTCTime)

The mailbox export job end timestamp.

describeMailboxExportJobResponse_s3Path :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The path to the S3 bucket and file that the mailbox export job is exporting to.

describeMailboxExportJobResponse_s3Prefix :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The S3 bucket prefix.

describeMailboxExportJobResponse_entityId :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The identifier of the user or resource associated with the mailbox.

describeMailboxExportJobResponse_description :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The mailbox export job description.

describeMailboxExportJobResponse_errorInfo :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

Error information for failed mailbox export jobs.

describeMailboxExportJobResponse_s3BucketName :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The name of the S3 bucket.

describeMailboxExportJobResponse_roleArn :: Lens' DescribeMailboxExportJobResponse (Maybe Text)

The ARN of the AWS Identity and Access Management (IAM) role that grants write permission to the Amazon Simple Storage Service (Amazon S3) bucket.

describeMailboxExportJobResponse_httpStatus :: Lens' DescribeMailboxExportJobResponse Int

The response's http status code.