Amazonka.WorkMail.DescribeResource

Description

Returns the data available for the resource.

Creating a Request

data DescribeResource

See: newDescribeResource smart constructor.

Constructors

DescribeResource'

Fields

organizationId :: Text

The identifier associated with the organization for which the resource is described.

resourceId :: Text

The identifier of the resource to be described.

newDescribeResource

Arguments

:: Text
$sel:organizationId:DescribeResource'

-> Text
$sel:resourceId:DescribeResource'

-> DescribeResource

Create a value of DescribeResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationId:DescribeResource', describeResource_organizationId - The identifier associated with the organization for which the resource is described.

$sel:resourceId:DescribeResource', describeResource_resourceId - The identifier of the resource to be described.

Request Lenses

describeResource_organizationId :: Lens' DescribeResource Text

The identifier associated with the organization for which the resource is described.

describeResource_resourceId :: Lens' DescribeResource Text

The identifier of the resource to be described.

Destructuring the Response

data DescribeResourceResponse

See: newDescribeResourceResponse smart constructor.

Constructors

DescribeResourceResponse'

Fields

email :: Maybe Text

The email of the described resource.

state :: Maybe EntityState

The state of the resource: enabled (registered to Amazon WorkMail), disabled (deregistered or never registered to WorkMail), or deleted.

resourceId :: Maybe Text

The identifier of the described resource.

disabledDate :: Maybe POSIX

The date and time when a resource was disabled from WorkMail, in UNIX epoch time format.

name :: Maybe Text

The name of the described resource.

type' :: Maybe ResourceType

The type of the described resource.

enabledDate :: Maybe POSIX

The date and time when a resource was enabled for WorkMail, in UNIX epoch time format.

bookingOptions :: Maybe BookingOptions

The booking options for the described resource.

httpStatus :: Int

The response's http status code.

newDescribeResourceResponse

Arguments

:: Int
$sel:httpStatus:DescribeResourceResponse'

-> DescribeResourceResponse

Create a value of DescribeResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:DescribeResourceResponse', describeResourceResponse_email - The email of the described resource.

$sel:state:DescribeResourceResponse', describeResourceResponse_state - The state of the resource: enabled (registered to Amazon WorkMail), disabled (deregistered or never registered to WorkMail), or deleted.

$sel:resourceId:DescribeResource', describeResourceResponse_resourceId - The identifier of the described resource.

$sel:disabledDate:DescribeResourceResponse', describeResourceResponse_disabledDate - The date and time when a resource was disabled from WorkMail, in UNIX epoch time format.

$sel:name:DescribeResourceResponse', describeResourceResponse_name - The name of the described resource.

$sel:type':DescribeResourceResponse', describeResourceResponse_type - The type of the described resource.

$sel:enabledDate:DescribeResourceResponse', describeResourceResponse_enabledDate - The date and time when a resource was enabled for WorkMail, in UNIX epoch time format.

$sel:bookingOptions:DescribeResourceResponse', describeResourceResponse_bookingOptions - The booking options for the described resource.

$sel:httpStatus:DescribeResourceResponse', describeResourceResponse_httpStatus - The response's http status code.

Response Lenses

describeResourceResponse_email :: Lens' DescribeResourceResponse (Maybe Text)

The email of the described resource.

describeResourceResponse_state :: Lens' DescribeResourceResponse (Maybe EntityState)

The state of the resource: enabled (registered to Amazon WorkMail), disabled (deregistered or never registered to WorkMail), or deleted.

describeResourceResponse_resourceId :: Lens' DescribeResourceResponse (Maybe Text)

The identifier of the described resource.

describeResourceResponse_disabledDate :: Lens' DescribeResourceResponse (Maybe UTCTime)

The date and time when a resource was disabled from WorkMail, in UNIX epoch time format.

describeResourceResponse_name :: Lens' DescribeResourceResponse (Maybe Text)

The name of the described resource.

describeResourceResponse_type :: Lens' DescribeResourceResponse (Maybe ResourceType)

The type of the described resource.

describeResourceResponse_enabledDate :: Lens' DescribeResourceResponse (Maybe UTCTime)

The date and time when a resource was enabled for WorkMail, in UNIX epoch time format.

describeResourceResponse_bookingOptions :: Lens' DescribeResourceResponse (Maybe BookingOptions)

The booking options for the described resource.

describeResourceResponse_httpStatus :: Lens' DescribeResourceResponse Int

The response's http status code.