Amazonka.WorkMail.DescribeUser

Description

Provides information regarding the user.

Creating a Request

data DescribeUser

See: newDescribeUser smart constructor.

Constructors

DescribeUser'

Fields

organizationId :: Text

The identifier for the organization under which the user exists.

userId :: Text

The identifier for the user to be described.

newDescribeUser

Arguments

:: Text
$sel:organizationId:DescribeUser'

-> Text
$sel:userId:DescribeUser'

-> DescribeUser

Create a value of DescribeUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationId:DescribeUser', describeUser_organizationId - The identifier for the organization under which the user exists.

$sel:userId:DescribeUser', describeUser_userId - The identifier for the user to be described.

Request Lenses

describeUser_organizationId :: Lens' DescribeUser Text

The identifier for the organization under which the user exists.

describeUser_userId :: Lens' DescribeUser Text

The identifier for the user to be described.

Destructuring the Response

data DescribeUserResponse

See: newDescribeUserResponse smart constructor.

Constructors

DescribeUserResponse'

Fields

email :: Maybe Text

The email of the user.

state :: Maybe EntityState

The state of a user: enabled (registered to Amazon WorkMail) or disabled (deregistered or never registered to WorkMail).

userId :: Maybe Text

The identifier for the described user.

disabledDate :: Maybe POSIX

The date and time at which the user was disabled for Amazon WorkMail usage, in UNIX epoch time format.

name :: Maybe Text

The name for the user.

displayName :: Maybe Text

The display name of the user.

userRole :: Maybe UserRole

In certain cases, other entities are modeled as users. If interoperability is enabled, resources are imported into Amazon WorkMail as users. Because different WorkMail organizations rely on different directory types, administrators can distinguish between an unregistered user (account is disabled and has a user role) and the directory administrators. The values are USER, RESOURCE, and SYSTEM_USER.

enabledDate :: Maybe POSIX

The date and time at which the user was enabled for Amazon WorkMail usage, in UNIX epoch time format.

httpStatus :: Int

The response's http status code.

newDescribeUserResponse

Arguments

:: Int
$sel:httpStatus:DescribeUserResponse'

-> DescribeUserResponse

Create a value of DescribeUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:DescribeUserResponse', describeUserResponse_email - The email of the user.

$sel:state:DescribeUserResponse', describeUserResponse_state - The state of a user: enabled (registered to Amazon WorkMail) or disabled (deregistered or never registered to WorkMail).

$sel:userId:DescribeUser', describeUserResponse_userId - The identifier for the described user.

$sel:disabledDate:DescribeUserResponse', describeUserResponse_disabledDate - The date and time at which the user was disabled for Amazon WorkMail usage, in UNIX epoch time format.

$sel:name:DescribeUserResponse', describeUserResponse_name - The name for the user.

$sel:displayName:DescribeUserResponse', describeUserResponse_displayName - The display name of the user.

$sel:userRole:DescribeUserResponse', describeUserResponse_userRole - In certain cases, other entities are modeled as users. If interoperability is enabled, resources are imported into Amazon WorkMail as users. Because different WorkMail organizations rely on different directory types, administrators can distinguish between an unregistered user (account is disabled and has a user role) and the directory administrators. The values are USER, RESOURCE, and SYSTEM_USER.

$sel:enabledDate:DescribeUserResponse', describeUserResponse_enabledDate - The date and time at which the user was enabled for Amazon WorkMail usage, in UNIX epoch time format.

$sel:httpStatus:DescribeUserResponse', describeUserResponse_httpStatus - The response's http status code.

Response Lenses

describeUserResponse_email :: Lens' DescribeUserResponse (Maybe Text)

The email of the user.

describeUserResponse_state :: Lens' DescribeUserResponse (Maybe EntityState)

The state of a user: enabled (registered to Amazon WorkMail) or disabled (deregistered or never registered to WorkMail).

describeUserResponse_userId :: Lens' DescribeUserResponse (Maybe Text)

The identifier for the described user.

describeUserResponse_disabledDate :: Lens' DescribeUserResponse (Maybe UTCTime)

The date and time at which the user was disabled for Amazon WorkMail usage, in UNIX epoch time format.

describeUserResponse_name :: Lens' DescribeUserResponse (Maybe Text)

The name for the user.

describeUserResponse_displayName :: Lens' DescribeUserResponse (Maybe Text)

The display name of the user.

describeUserResponse_userRole :: Lens' DescribeUserResponse (Maybe UserRole)

In certain cases, other entities are modeled as users. If interoperability is enabled, resources are imported into Amazon WorkMail as users. Because different WorkMail organizations rely on different directory types, administrators can distinguish between an unregistered user (account is disabled and has a user role) and the directory administrators. The values are USER, RESOURCE, and SYSTEM_USER.

describeUserResponse_enabledDate :: Lens' DescribeUserResponse (Maybe UTCTime)

The date and time at which the user was enabled for Amazon WorkMail usage, in UNIX epoch time format.

describeUserResponse_httpStatus :: Lens' DescribeUserResponse Int

The response's http status code.