Amazonka.WorkMail.DisassociateMemberFromGroup

Description

Removes a member from a group.

Creating a Request

data DisassociateMemberFromGroup

See: newDisassociateMemberFromGroup smart constructor.

Constructors

DisassociateMemberFromGroup'

Fields

organizationId :: Text

The identifier for the organization under which the group exists.

groupId :: Text

The identifier for the group from which members are removed.

memberId :: Text

The identifier for the member to be removed to the group.

newDisassociateMemberFromGroup

Arguments

:: Text
$sel:organizationId:DisassociateMemberFromGroup'

-> Text
$sel:groupId:DisassociateMemberFromGroup'

-> Text
$sel:memberId:DisassociateMemberFromGroup'

-> DisassociateMemberFromGroup

Create a value of DisassociateMemberFromGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organizationId:DisassociateMemberFromGroup', disassociateMemberFromGroup_organizationId - The identifier for the organization under which the group exists.

$sel:groupId:DisassociateMemberFromGroup', disassociateMemberFromGroup_groupId - The identifier for the group from which members are removed.

$sel:memberId:DisassociateMemberFromGroup', disassociateMemberFromGroup_memberId - The identifier for the member to be removed to the group.

Request Lenses

disassociateMemberFromGroup_organizationId :: Lens' DisassociateMemberFromGroup Text

The identifier for the organization under which the group exists.

disassociateMemberFromGroup_groupId :: Lens' DisassociateMemberFromGroup Text

The identifier for the group from which members are removed.

disassociateMemberFromGroup_memberId :: Lens' DisassociateMemberFromGroup Text

The identifier for the member to be removed to the group.

Destructuring the Response

data DisassociateMemberFromGroupResponse

See: newDisassociateMemberFromGroupResponse smart constructor.

Constructors

DisassociateMemberFromGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

newDisassociateMemberFromGroupResponse

Arguments

:: Int
$sel:httpStatus:DisassociateMemberFromGroupResponse'

-> DisassociateMemberFromGroupResponse

Create a value of DisassociateMemberFromGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateMemberFromGroupResponse', disassociateMemberFromGroupResponse_httpStatus - The response's http status code.

Response Lenses

disassociateMemberFromGroupResponse_httpStatus :: Lens' DisassociateMemberFromGroupResponse Int

The response's http status code.