Amazonka.WorkMail.GetMobileDeviceAccessEffect

Description

Simulates the effect of the mobile device access rules for the given attributes of a sample access event. Use this method to test the effects of the current set of mobile device access rules for the Amazon WorkMail organization for a particular user's attributes.

Creating a Request

data GetMobileDeviceAccessEffect

See: newGetMobileDeviceAccessEffect smart constructor.

Constructors

GetMobileDeviceAccessEffect'

Fields

deviceOperatingSystem :: Maybe Text

Device operating system the simulated user will report.

deviceModel :: Maybe Text

Device model the simulated user will report.

deviceType :: Maybe Text

Device type the simulated user will report.

deviceUserAgent :: Maybe Text

Device user agent the simulated user will report.

organizationId :: Text

The Amazon WorkMail organization to simulate the access effect for.

newGetMobileDeviceAccessEffect

Arguments

:: Text
$sel:organizationId:GetMobileDeviceAccessEffect'

-> GetMobileDeviceAccessEffect

Create a value of GetMobileDeviceAccessEffect with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceOperatingSystem:GetMobileDeviceAccessEffect', getMobileDeviceAccessEffect_deviceOperatingSystem - Device operating system the simulated user will report.

$sel:deviceModel:GetMobileDeviceAccessEffect', getMobileDeviceAccessEffect_deviceModel - Device model the simulated user will report.

$sel:deviceType:GetMobileDeviceAccessEffect', getMobileDeviceAccessEffect_deviceType - Device type the simulated user will report.

$sel:deviceUserAgent:GetMobileDeviceAccessEffect', getMobileDeviceAccessEffect_deviceUserAgent - Device user agent the simulated user will report.

$sel:organizationId:GetMobileDeviceAccessEffect', getMobileDeviceAccessEffect_organizationId - The Amazon WorkMail organization to simulate the access effect for.

Request Lenses

getMobileDeviceAccessEffect_deviceOperatingSystem :: Lens' GetMobileDeviceAccessEffect (Maybe Text)

Device operating system the simulated user will report.

getMobileDeviceAccessEffect_deviceModel :: Lens' GetMobileDeviceAccessEffect (Maybe Text)

Device model the simulated user will report.

getMobileDeviceAccessEffect_deviceType :: Lens' GetMobileDeviceAccessEffect (Maybe Text)

Device type the simulated user will report.

getMobileDeviceAccessEffect_deviceUserAgent :: Lens' GetMobileDeviceAccessEffect (Maybe Text)

Device user agent the simulated user will report.

getMobileDeviceAccessEffect_organizationId :: Lens' GetMobileDeviceAccessEffect Text

The Amazon WorkMail organization to simulate the access effect for.

Destructuring the Response

data GetMobileDeviceAccessEffectResponse

See: newGetMobileDeviceAccessEffectResponse smart constructor.

Constructors

GetMobileDeviceAccessEffectResponse'

Fields

effect :: Maybe MobileDeviceAccessRuleEffect

The effect of the simulated access, ALLOW or DENY, after evaluating mobile device access rules in the Amazon WorkMail organization for the simulated user parameters.

matchedRules :: Maybe [MobileDeviceAccessMatchedRule]

A list of the rules which matched the simulated user input and produced the effect.

httpStatus :: Int

The response's http status code.

newGetMobileDeviceAccessEffectResponse

Arguments

:: Int
$sel:httpStatus:GetMobileDeviceAccessEffectResponse'

-> GetMobileDeviceAccessEffectResponse

Create a value of GetMobileDeviceAccessEffectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:effect:GetMobileDeviceAccessEffectResponse', getMobileDeviceAccessEffectResponse_effect - The effect of the simulated access, ALLOW or DENY, after evaluating mobile device access rules in the Amazon WorkMail organization for the simulated user parameters.

$sel:matchedRules:GetMobileDeviceAccessEffectResponse', getMobileDeviceAccessEffectResponse_matchedRules - A list of the rules which matched the simulated user input and produced the effect.

$sel:httpStatus:GetMobileDeviceAccessEffectResponse', getMobileDeviceAccessEffectResponse_httpStatus - The response's http status code.

Response Lenses

getMobileDeviceAccessEffectResponse_effect :: Lens' GetMobileDeviceAccessEffectResponse (Maybe MobileDeviceAccessRuleEffect)

The effect of the simulated access, ALLOW or DENY, after evaluating mobile device access rules in the Amazon WorkMail organization for the simulated user parameters.

getMobileDeviceAccessEffectResponse_matchedRules :: Lens' GetMobileDeviceAccessEffectResponse (Maybe [MobileDeviceAccessMatchedRule])

A list of the rules which matched the simulated user input and produced the effect.

getMobileDeviceAccessEffectResponse_httpStatus :: Lens' GetMobileDeviceAccessEffectResponse Int

The response's http status code.