Amazonka.WorkMail.ListGroupMembers

Description

Returns an overview of the members of a group. Users and groups can be members of a group.

This operation returns paginated results.

Creating a Request

data ListGroupMembers

See: newListGroupMembers smart constructor.

Constructors

ListGroupMembers'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. The first call does not contain any tokens.

maxResults :: Maybe Natural

The maximum number of results to return in a single call.

organizationId :: Text

The identifier for the organization under which the group exists.

groupId :: Text

The identifier for the group to which the members (users or groups) are associated.

newListGroupMembers

Arguments

:: Text
$sel:organizationId:ListGroupMembers'

-> Text
$sel:groupId:ListGroupMembers'

-> ListGroupMembers

Create a value of ListGroupMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGroupMembers', listGroupMembers_nextToken - The token to use to retrieve the next page of results. The first call does not contain any tokens.

$sel:maxResults:ListGroupMembers', listGroupMembers_maxResults - The maximum number of results to return in a single call.

$sel:organizationId:ListGroupMembers', listGroupMembers_organizationId - The identifier for the organization under which the group exists.

$sel:groupId:ListGroupMembers', listGroupMembers_groupId - The identifier for the group to which the members (users or groups) are associated.

Request Lenses

listGroupMembers_nextToken :: Lens' ListGroupMembers (Maybe Text)

The token to use to retrieve the next page of results. The first call does not contain any tokens.

listGroupMembers_maxResults :: Lens' ListGroupMembers (Maybe Natural)

The maximum number of results to return in a single call.

listGroupMembers_organizationId :: Lens' ListGroupMembers Text

The identifier for the organization under which the group exists.

listGroupMembers_groupId :: Lens' ListGroupMembers Text

The identifier for the group to which the members (users or groups) are associated.

Destructuring the Response

data ListGroupMembersResponse

See: newListGroupMembersResponse smart constructor.

Constructors

ListGroupMembersResponse'

Fields

members :: Maybe [Member]

The members associated to the group.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. The first call does not contain any tokens.

httpStatus :: Int

The response's http status code.

newListGroupMembersResponse

Arguments

:: Int
$sel:httpStatus:ListGroupMembersResponse'

-> ListGroupMembersResponse

Create a value of ListGroupMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:members:ListGroupMembersResponse', listGroupMembersResponse_members - The members associated to the group.

$sel:nextToken:ListGroupMembers', listGroupMembersResponse_nextToken - The token to use to retrieve the next page of results. The first call does not contain any tokens.

$sel:httpStatus:ListGroupMembersResponse', listGroupMembersResponse_httpStatus - The response's http status code.

Response Lenses

listGroupMembersResponse_members :: Lens' ListGroupMembersResponse (Maybe [Member])

The members associated to the group.

listGroupMembersResponse_nextToken :: Lens' ListGroupMembersResponse (Maybe Text)

The token to use to retrieve the next page of results. The first call does not contain any tokens.

listGroupMembersResponse_httpStatus :: Lens' ListGroupMembersResponse Int

The response's http status code.