Amazonka.WorkMail.ListMailboxPermissions

Description

Lists the mailbox permissions associated with a user, group, or resource mailbox.

This operation returns paginated results.

Creating a Request

data ListMailboxPermissions

See: newListMailboxPermissions smart constructor.

Constructors

ListMailboxPermissions'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. The first call does not contain any tokens.

maxResults :: Maybe Natural

The maximum number of results to return in a single call.

organizationId :: Text

The identifier of the organization under which the user, group, or resource exists.

entityId :: Text

The identifier of the user, group, or resource for which to list mailbox permissions.

newListMailboxPermissions

Arguments

:: Text
$sel:organizationId:ListMailboxPermissions'

-> Text
$sel:entityId:ListMailboxPermissions'

-> ListMailboxPermissions

Create a value of ListMailboxPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMailboxPermissions', listMailboxPermissions_nextToken - The token to use to retrieve the next page of results. The first call does not contain any tokens.

$sel:maxResults:ListMailboxPermissions', listMailboxPermissions_maxResults - The maximum number of results to return in a single call.

$sel:organizationId:ListMailboxPermissions', listMailboxPermissions_organizationId - The identifier of the organization under which the user, group, or resource exists.

$sel:entityId:ListMailboxPermissions', listMailboxPermissions_entityId - The identifier of the user, group, or resource for which to list mailbox permissions.

Request Lenses

listMailboxPermissions_nextToken :: Lens' ListMailboxPermissions (Maybe Text)

The token to use to retrieve the next page of results. The first call does not contain any tokens.

listMailboxPermissions_maxResults :: Lens' ListMailboxPermissions (Maybe Natural)

The maximum number of results to return in a single call.

listMailboxPermissions_organizationId :: Lens' ListMailboxPermissions Text

The identifier of the organization under which the user, group, or resource exists.

listMailboxPermissions_entityId :: Lens' ListMailboxPermissions Text

The identifier of the user, group, or resource for which to list mailbox permissions.

Destructuring the Response

data ListMailboxPermissionsResponse

See: newListMailboxPermissionsResponse smart constructor.

Constructors

ListMailboxPermissionsResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. The value is "null" when there are no more results to return.

permissions :: Maybe [Permission]

One page of the user, group, or resource mailbox permissions.

httpStatus :: Int

The response's http status code.

newListMailboxPermissionsResponse

Arguments

:: Int
$sel:httpStatus:ListMailboxPermissionsResponse'

-> ListMailboxPermissionsResponse

Create a value of ListMailboxPermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMailboxPermissions', listMailboxPermissionsResponse_nextToken - The token to use to retrieve the next page of results. The value is "null" when there are no more results to return.

$sel:permissions:ListMailboxPermissionsResponse', listMailboxPermissionsResponse_permissions - One page of the user, group, or resource mailbox permissions.

$sel:httpStatus:ListMailboxPermissionsResponse', listMailboxPermissionsResponse_httpStatus - The response's http status code.

Response Lenses

listMailboxPermissionsResponse_nextToken :: Lens' ListMailboxPermissionsResponse (Maybe Text)

The token to use to retrieve the next page of results. The value is "null" when there are no more results to return.

listMailboxPermissionsResponse_permissions :: Lens' ListMailboxPermissionsResponse (Maybe [Permission])

One page of the user, group, or resource mailbox permissions.

listMailboxPermissionsResponse_httpStatus :: Lens' ListMailboxPermissionsResponse Int

The response's http status code.