Amazonka.WorkMail.ListMobileDeviceAccessOverrides

Description

Lists all the mobile device access overrides for any given combination of WorkMail organization, user, or device.

Creating a Request

data ListMobileDeviceAccessOverrides

See: newListMobileDeviceAccessOverrides smart constructor.

Constructors

ListMobileDeviceAccessOverrides'

Fields

userId :: Maybe Text

The WorkMail user under which you list the mobile device access overrides. Accepts the following types of user identities:

User ID: 12345678-1234-1234-1234-123456789012 or S-1-1-12-1234567890-123456789-123456789-1234

Email address: user@domain.tld

User name: user

nextToken :: Maybe Text

The token to use to retrieve the next page of results. The first call does not require a token.

deviceId :: Maybe Text

The mobile device to which the access override applies.

maxResults :: Maybe Natural

The maximum number of results to return in a single call.

organizationId :: Text

The Amazon WorkMail organization under which to list mobile device access overrides.

newListMobileDeviceAccessOverrides

Arguments

:: Text
$sel:organizationId:ListMobileDeviceAccessOverrides'

-> ListMobileDeviceAccessOverrides

Create a value of ListMobileDeviceAccessOverrides with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userId:ListMobileDeviceAccessOverrides', listMobileDeviceAccessOverrides_userId - The WorkMail user under which you list the mobile device access overrides. Accepts the following types of user identities:

User ID: 12345678-1234-1234-1234-123456789012 or S-1-1-12-1234567890-123456789-123456789-1234

Email address: user@domain.tld

User name: user

$sel:nextToken:ListMobileDeviceAccessOverrides', listMobileDeviceAccessOverrides_nextToken - The token to use to retrieve the next page of results. The first call does not require a token.

$sel:deviceId:ListMobileDeviceAccessOverrides', listMobileDeviceAccessOverrides_deviceId - The mobile device to which the access override applies.

$sel:maxResults:ListMobileDeviceAccessOverrides', listMobileDeviceAccessOverrides_maxResults - The maximum number of results to return in a single call.

$sel:organizationId:ListMobileDeviceAccessOverrides', listMobileDeviceAccessOverrides_organizationId - The Amazon WorkMail organization under which to list mobile device access overrides.

Request Lenses

listMobileDeviceAccessOverrides_userId :: Lens' ListMobileDeviceAccessOverrides (Maybe Text)

The WorkMail user under which you list the mobile device access overrides. Accepts the following types of user identities:

User ID: 12345678-1234-1234-1234-123456789012 or S-1-1-12-1234567890-123456789-123456789-1234

Email address: user@domain.tld

User name: user

listMobileDeviceAccessOverrides_nextToken :: Lens' ListMobileDeviceAccessOverrides (Maybe Text)

The token to use to retrieve the next page of results. The first call does not require a token.

listMobileDeviceAccessOverrides_deviceId :: Lens' ListMobileDeviceAccessOverrides (Maybe Text)

The mobile device to which the access override applies.

listMobileDeviceAccessOverrides_maxResults :: Lens' ListMobileDeviceAccessOverrides (Maybe Natural)

The maximum number of results to return in a single call.

listMobileDeviceAccessOverrides_organizationId :: Lens' ListMobileDeviceAccessOverrides Text

The Amazon WorkMail organization under which to list mobile device access overrides.

Destructuring the Response

data ListMobileDeviceAccessOverridesResponse

See: newListMobileDeviceAccessOverridesResponse smart constructor.

Constructors

ListMobileDeviceAccessOverridesResponse'

Fields

overrides :: Maybe [MobileDeviceAccessOverride]

The list of mobile device access overrides that exist for the specified Amazon WorkMail organization and user.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. The value is “null” when there are no more results to return.

httpStatus :: Int

The response's http status code.

newListMobileDeviceAccessOverridesResponse

Arguments

:: Int
$sel:httpStatus:ListMobileDeviceAccessOverridesResponse'

-> ListMobileDeviceAccessOverridesResponse

Create a value of ListMobileDeviceAccessOverridesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:overrides:ListMobileDeviceAccessOverridesResponse', listMobileDeviceAccessOverridesResponse_overrides - The list of mobile device access overrides that exist for the specified Amazon WorkMail organization and user.

$sel:nextToken:ListMobileDeviceAccessOverrides', listMobileDeviceAccessOverridesResponse_nextToken - The token to use to retrieve the next page of results. The value is “null” when there are no more results to return.

$sel:httpStatus:ListMobileDeviceAccessOverridesResponse', listMobileDeviceAccessOverridesResponse_httpStatus - The response's http status code.

Response Lenses

listMobileDeviceAccessOverridesResponse_overrides :: Lens' ListMobileDeviceAccessOverridesResponse (Maybe [MobileDeviceAccessOverride])

The list of mobile device access overrides that exist for the specified Amazon WorkMail organization and user.

listMobileDeviceAccessOverridesResponse_nextToken :: Lens' ListMobileDeviceAccessOverridesResponse (Maybe Text)

The token to use to retrieve the next page of results. The value is “null” when there are no more results to return.

listMobileDeviceAccessOverridesResponse_httpStatus :: Lens' ListMobileDeviceAccessOverridesResponse Int

The response's http status code.