Amazonka.WorkMail.PutMobileDeviceAccessOverride

Description

Creates or updates a mobile device access override for the given WorkMail organization, user, and device.

Creating a Request

data PutMobileDeviceAccessOverride

See: newPutMobileDeviceAccessOverride smart constructor.

Constructors

PutMobileDeviceAccessOverride'

Fields

description :: Maybe Text

A description of the override.

organizationId :: Text

Identifies the Amazon WorkMail organization for which you create the override.

userId :: Text

The WorkMail user for which you create the override. Accepts the following types of user identities:

User ID: 12345678-1234-1234-1234-123456789012 or S-1-1-12-1234567890-123456789-123456789-1234

Email address: user@domain.tld

User name: user

deviceId :: Text

The mobile device for which you create the override. DeviceId is case insensitive.

effect :: MobileDeviceAccessRuleEffect

The effect of the override, ALLOW or DENY.

newPutMobileDeviceAccessOverride

Arguments

:: Text
$sel:organizationId:PutMobileDeviceAccessOverride'

-> Text
$sel:userId:PutMobileDeviceAccessOverride'

-> Text
$sel:deviceId:PutMobileDeviceAccessOverride'

-> MobileDeviceAccessRuleEffect
$sel:effect:PutMobileDeviceAccessOverride'

-> PutMobileDeviceAccessOverride

Create a value of PutMobileDeviceAccessOverride with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:PutMobileDeviceAccessOverride', putMobileDeviceAccessOverride_description - A description of the override.

$sel:organizationId:PutMobileDeviceAccessOverride', putMobileDeviceAccessOverride_organizationId - Identifies the Amazon WorkMail organization for which you create the override.

$sel:userId:PutMobileDeviceAccessOverride', putMobileDeviceAccessOverride_userId - The WorkMail user for which you create the override. Accepts the following types of user identities:

User ID: 12345678-1234-1234-1234-123456789012 or S-1-1-12-1234567890-123456789-123456789-1234

Email address: user@domain.tld

User name: user

$sel:deviceId:PutMobileDeviceAccessOverride', putMobileDeviceAccessOverride_deviceId - The mobile device for which you create the override. DeviceId is case insensitive.

$sel:effect:PutMobileDeviceAccessOverride', putMobileDeviceAccessOverride_effect - The effect of the override, ALLOW or DENY.

Request Lenses

putMobileDeviceAccessOverride_description :: Lens' PutMobileDeviceAccessOverride (Maybe Text)

A description of the override.

putMobileDeviceAccessOverride_organizationId :: Lens' PutMobileDeviceAccessOverride Text

Identifies the Amazon WorkMail organization for which you create the override.

putMobileDeviceAccessOverride_userId :: Lens' PutMobileDeviceAccessOverride Text

The WorkMail user for which you create the override. Accepts the following types of user identities:

User ID: 12345678-1234-1234-1234-123456789012 or S-1-1-12-1234567890-123456789-123456789-1234

Email address: user@domain.tld

User name: user

putMobileDeviceAccessOverride_deviceId :: Lens' PutMobileDeviceAccessOverride Text

The mobile device for which you create the override. DeviceId is case insensitive.

putMobileDeviceAccessOverride_effect :: Lens' PutMobileDeviceAccessOverride MobileDeviceAccessRuleEffect

The effect of the override, ALLOW or DENY.

Destructuring the Response

data PutMobileDeviceAccessOverrideResponse

See: newPutMobileDeviceAccessOverrideResponse smart constructor.

Constructors

PutMobileDeviceAccessOverrideResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutMobileDeviceAccessOverrideResponse

Arguments

:: Int
$sel:httpStatus:PutMobileDeviceAccessOverrideResponse'

-> PutMobileDeviceAccessOverrideResponse

Create a value of PutMobileDeviceAccessOverrideResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutMobileDeviceAccessOverrideResponse', putMobileDeviceAccessOverrideResponse_httpStatus - The response's http status code.

Response Lenses

putMobileDeviceAccessOverrideResponse_httpStatus :: Lens' PutMobileDeviceAccessOverrideResponse Int

The response's http status code.