Amazonka.WorkMail.Types.AccessControlRule

Description

Documentation

data AccessControlRule

A rule that controls access to an Amazon WorkMail organization.

See: newAccessControlRule smart constructor.

Constructors

AccessControlRule'

Fields

effect :: Maybe AccessControlRuleEffect

The rule effect.

userIds :: Maybe [Text]

User IDs to include in the rule.

actions :: Maybe [Text]

Access protocol actions to include in the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

dateCreated :: Maybe POSIX

The date that the rule was created.

name :: Maybe Text

The rule name.

notUserIds :: Maybe [Text]

User IDs to exclude from the rule.

dateModified :: Maybe POSIX

The date that the rule was modified.

ipRanges :: Maybe [Text]

IPv4 CIDR ranges to include in the rule.

notIpRanges :: Maybe [Text]

IPv4 CIDR ranges to exclude from the rule.

notActions :: Maybe [Text]

Access protocol actions to exclude from the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

description :: Maybe Text

The rule description.

newAccessControlRule :: AccessControlRule

Create a value of AccessControlRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:effect:AccessControlRule', accessControlRule_effect - The rule effect.

$sel:userIds:AccessControlRule', accessControlRule_userIds - User IDs to include in the rule.

$sel:actions:AccessControlRule', accessControlRule_actions - Access protocol actions to include in the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

$sel:dateCreated:AccessControlRule', accessControlRule_dateCreated - The date that the rule was created.

$sel:name:AccessControlRule', accessControlRule_name - The rule name.

$sel:notUserIds:AccessControlRule', accessControlRule_notUserIds - User IDs to exclude from the rule.

$sel:dateModified:AccessControlRule', accessControlRule_dateModified - The date that the rule was modified.

$sel:ipRanges:AccessControlRule', accessControlRule_ipRanges - IPv4 CIDR ranges to include in the rule.

$sel:notIpRanges:AccessControlRule', accessControlRule_notIpRanges - IPv4 CIDR ranges to exclude from the rule.

$sel:notActions:AccessControlRule', accessControlRule_notActions - Access protocol actions to exclude from the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

$sel:description:AccessControlRule', accessControlRule_description - The rule description.

accessControlRule_effect :: Lens' AccessControlRule (Maybe AccessControlRuleEffect)

The rule effect.

accessControlRule_userIds :: Lens' AccessControlRule (Maybe [Text])

User IDs to include in the rule.

accessControlRule_actions :: Lens' AccessControlRule (Maybe [Text])

Access protocol actions to include in the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

accessControlRule_dateCreated :: Lens' AccessControlRule (Maybe UTCTime)

The date that the rule was created.

accessControlRule_name :: Lens' AccessControlRule (Maybe Text)

The rule name.

accessControlRule_notUserIds :: Lens' AccessControlRule (Maybe [Text])

User IDs to exclude from the rule.

accessControlRule_dateModified :: Lens' AccessControlRule (Maybe UTCTime)

The date that the rule was modified.

accessControlRule_ipRanges :: Lens' AccessControlRule (Maybe [Text])

IPv4 CIDR ranges to include in the rule.

accessControlRule_notIpRanges :: Lens' AccessControlRule (Maybe [Text])

IPv4 CIDR ranges to exclude from the rule.

accessControlRule_notActions :: Lens' AccessControlRule (Maybe [Text])

Access protocol actions to exclude from the rule. Valid values include ActiveSync, AutoDiscover, EWS, IMAP, SMTP, WindowsOutlook, and WebMail.

accessControlRule_description :: Lens' AccessControlRule (Maybe Text)

The rule description.