Amazonka.WorkMailMessageFlow.Types

Description

Service Configuration

defaultService :: Service

API version 2019-05-01 of the Amazon WorkMail Message Flow SDK configuration.

Errors

_MessageRejected :: AsError a => Getting (First ServiceError) a ServiceError

The requested email could not be updated due to an error in the MIME content. Check the error message for more information about what caused the error.

_InvalidContentLocation :: AsError a => Getting (First ServiceError) a ServiceError

WorkMail could not access the updated email content. Possible reasons:

You made the request in a region other than your S3 bucket region.

The S3 bucket owner is not the same as the calling AWS account.

You have an incomplete or missing S3 bucket policy. For more information about policies, see Updating message content with AWS Lambda in the WorkMail Administrator Guide.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested email message is not found.

_MessageFrozen :: AsError a => Getting (First ServiceError) a ServiceError

The requested email is not eligible for update. This is usually the case for a redirected email.

RawMessageContent

data RawMessageContent

Provides the MIME content of the updated email message as an S3 object. All MIME content must meet the following criteria:

Each part of a multipart MIME message must be formatted properly.

Attachments must be of a content type that Amazon SES supports. For more information, see Unsupported Attachment Types.

If any of the MIME parts in a message contain content that is outside of the 7-bit ASCII character range, we recommend encoding that content.

Per RFC 5321, the maximum length of each line of text, including the <CRLF>, must not exceed 1,000 characters.

The message must contain all the required header fields. Check the returned error message for more information.

The value of immutable headers must remain unchanged. Check the returned error message for more information.

Certain unique headers can only appear once. Check the returned error message for more information.

See: newRawMessageContent smart constructor.

Constructors

RawMessageContent'

Fields

s3Reference :: S3Reference

The S3 reference of an email message.

newRawMessageContent

Arguments

:: S3Reference
$sel:s3Reference:RawMessageContent'

-> RawMessageContent

Create a value of RawMessageContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Reference:RawMessageContent', rawMessageContent_s3Reference - The S3 reference of an email message.

rawMessageContent_s3Reference :: Lens' RawMessageContent S3Reference

The S3 reference of an email message.

S3Reference

data S3Reference

Amazon S3 object representing the updated message content, in MIME format.

The region for the S3 bucket containing the S3 object must match the region used for WorkMail operations. Also, for WorkMail to process an S3 object, it must have permission to access that object. For more information, see Updating message content with AWS Lambda.

See: newS3Reference smart constructor.

Constructors

S3Reference'

Fields

objectVersion :: Maybe Text

If you enable versioning for the bucket, you can specify the object version.

bucket :: Text

The S3 bucket name.

key :: Text

The S3 key object name.

newS3Reference

Arguments

:: Text
$sel:bucket:S3Reference'

-> Text
$sel:key:S3Reference'

-> S3Reference

Create a value of S3Reference with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:objectVersion:S3Reference', s3Reference_objectVersion - If you enable versioning for the bucket, you can specify the object version.

$sel:bucket:S3Reference', s3Reference_bucket - The S3 bucket name.

$sel:key:S3Reference', s3Reference_key - The S3 key object name.

s3Reference_objectVersion :: Lens' S3Reference (Maybe Text)

If you enable versioning for the bucket, you can specify the object version.

s3Reference_bucket :: Lens' S3Reference Text

The S3 bucket name.

s3Reference_key :: Lens' S3Reference Text

The S3 key object name.