Amazonka.WorkMailMessageFlow.Types.RawMessageContent

Description

Documentation

data RawMessageContent

Provides the MIME content of the updated email message as an S3 object. All MIME content must meet the following criteria:

Each part of a multipart MIME message must be formatted properly.

Attachments must be of a content type that Amazon SES supports. For more information, see Unsupported Attachment Types.

If any of the MIME parts in a message contain content that is outside of the 7-bit ASCII character range, we recommend encoding that content.

Per RFC 5321, the maximum length of each line of text, including the <CRLF>, must not exceed 1,000 characters.

The message must contain all the required header fields. Check the returned error message for more information.

The value of immutable headers must remain unchanged. Check the returned error message for more information.

Certain unique headers can only appear once. Check the returned error message for more information.

See: newRawMessageContent smart constructor.

Constructors

RawMessageContent'

Fields

s3Reference :: S3Reference

The S3 reference of an email message.

newRawMessageContent

Arguments

:: S3Reference
$sel:s3Reference:RawMessageContent'

-> RawMessageContent

Create a value of RawMessageContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Reference:RawMessageContent', rawMessageContent_s3Reference - The S3 reference of an email message.

rawMessageContent_s3Reference :: Lens' RawMessageContent S3Reference

The S3 reference of an email message.