Amazonka.WorkSpaces.CreateConnectionAlias

Description

Creates the specified connection alias for use with cross-Region redirection. For more information, see Cross-Region Redirection for Amazon WorkSpaces.

Creating a Request

data CreateConnectionAlias

See: newCreateConnectionAlias smart constructor.

Constructors

CreateConnectionAlias'

Fields

tags :: Maybe [Tag]

The tags to associate with the connection alias.

connectionString :: Text

A connection string in the form of a fully qualified domain name (FQDN), such as www.example.com.

After you create a connection string, it is always associated to your Amazon Web Services account. You cannot recreate the same connection string with a different account, even if you delete all instances of it from the original account. The connection string is globally reserved for your account.

newCreateConnectionAlias

Arguments

:: Text
$sel:connectionString:CreateConnectionAlias'

-> CreateConnectionAlias

Create a value of CreateConnectionAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateConnectionAlias', createConnectionAlias_tags - The tags to associate with the connection alias.

$sel:connectionString:CreateConnectionAlias', createConnectionAlias_connectionString - A connection string in the form of a fully qualified domain name (FQDN), such as www.example.com.

After you create a connection string, it is always associated to your Amazon Web Services account. You cannot recreate the same connection string with a different account, even if you delete all instances of it from the original account. The connection string is globally reserved for your account.

Request Lenses

createConnectionAlias_tags :: Lens' CreateConnectionAlias (Maybe [Tag])

The tags to associate with the connection alias.

createConnectionAlias_connectionString :: Lens' CreateConnectionAlias Text

A connection string in the form of a fully qualified domain name (FQDN), such as www.example.com.

After you create a connection string, it is always associated to your Amazon Web Services account. You cannot recreate the same connection string with a different account, even if you delete all instances of it from the original account. The connection string is globally reserved for your account.

Destructuring the Response

data CreateConnectionAliasResponse

See: newCreateConnectionAliasResponse smart constructor.

Constructors

CreateConnectionAliasResponse'

Fields

aliasId :: Maybe Text

The identifier of the connection alias.

httpStatus :: Int

The response's http status code.

newCreateConnectionAliasResponse

Arguments

:: Int
$sel:httpStatus:CreateConnectionAliasResponse'

-> CreateConnectionAliasResponse

Create a value of CreateConnectionAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aliasId:CreateConnectionAliasResponse', createConnectionAliasResponse_aliasId - The identifier of the connection alias.

$sel:httpStatus:CreateConnectionAliasResponse', createConnectionAliasResponse_httpStatus - The response's http status code.

Response Lenses

createConnectionAliasResponse_aliasId :: Lens' CreateConnectionAliasResponse (Maybe Text)

The identifier of the connection alias.

createConnectionAliasResponse_httpStatus :: Lens' CreateConnectionAliasResponse Int

The response's http status code.