Amazonka.WorkSpaces.CreateWorkspaceBundle

Description

Creates the specified WorkSpace bundle. For more information about creating WorkSpace bundles, see Create a Custom WorkSpaces Image and Bundle.

Creating a Request

data CreateWorkspaceBundle

See: newCreateWorkspaceBundle smart constructor.

Constructors

CreateWorkspaceBundle'

Fields

rootStorage :: Maybe RootStorage

tags :: Maybe [Tag]

The tags associated with the bundle.

To add tags at the same time when you're creating the bundle, you must create an IAM policy that grants your IAM user permissions to use workspaces:CreateTags.

bundleName :: Text

The name of the bundle.

bundleDescription :: Text

The description of the bundle.

imageId :: Text

The identifier of the image that is used to create the bundle.

computeType :: ComputeType

userStorage :: UserStorage

newCreateWorkspaceBundle

Arguments

:: Text
$sel:bundleName:CreateWorkspaceBundle'

-> Text
$sel:bundleDescription:CreateWorkspaceBundle'

-> Text
$sel:imageId:CreateWorkspaceBundle'

-> ComputeType
$sel:computeType:CreateWorkspaceBundle'

-> UserStorage
$sel:userStorage:CreateWorkspaceBundle'

-> CreateWorkspaceBundle

Create a value of CreateWorkspaceBundle with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rootStorage:CreateWorkspaceBundle', createWorkspaceBundle_rootStorage - Undocumented member.

$sel:tags:CreateWorkspaceBundle', createWorkspaceBundle_tags - The tags associated with the bundle.

To add tags at the same time when you're creating the bundle, you must create an IAM policy that grants your IAM user permissions to use workspaces:CreateTags.

$sel:bundleName:CreateWorkspaceBundle', createWorkspaceBundle_bundleName - The name of the bundle.

$sel:bundleDescription:CreateWorkspaceBundle', createWorkspaceBundle_bundleDescription - The description of the bundle.

$sel:imageId:CreateWorkspaceBundle', createWorkspaceBundle_imageId - The identifier of the image that is used to create the bundle.

$sel:computeType:CreateWorkspaceBundle', createWorkspaceBundle_computeType - Undocumented member.

$sel:userStorage:CreateWorkspaceBundle', createWorkspaceBundle_userStorage - Undocumented member.

Request Lenses

createWorkspaceBundle_rootStorage :: Lens' CreateWorkspaceBundle (Maybe RootStorage)

Undocumented member.

createWorkspaceBundle_tags :: Lens' CreateWorkspaceBundle (Maybe [Tag])

The tags associated with the bundle.

To add tags at the same time when you're creating the bundle, you must create an IAM policy that grants your IAM user permissions to use workspaces:CreateTags.

createWorkspaceBundle_bundleName :: Lens' CreateWorkspaceBundle Text

The name of the bundle.

createWorkspaceBundle_bundleDescription :: Lens' CreateWorkspaceBundle Text

The description of the bundle.

createWorkspaceBundle_imageId :: Lens' CreateWorkspaceBundle Text

The identifier of the image that is used to create the bundle.

createWorkspaceBundle_computeType :: Lens' CreateWorkspaceBundle ComputeType

Undocumented member.

createWorkspaceBundle_userStorage :: Lens' CreateWorkspaceBundle UserStorage

Undocumented member.

Destructuring the Response

data CreateWorkspaceBundleResponse

See: newCreateWorkspaceBundleResponse smart constructor.

Constructors

CreateWorkspaceBundleResponse'

Fields

workspaceBundle :: Maybe WorkspaceBundle

httpStatus :: Int

The response's http status code.

newCreateWorkspaceBundleResponse

Arguments

:: Int
$sel:httpStatus:CreateWorkspaceBundleResponse'

-> CreateWorkspaceBundleResponse

Create a value of CreateWorkspaceBundleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceBundle:CreateWorkspaceBundleResponse', createWorkspaceBundleResponse_workspaceBundle - Undocumented member.

$sel:httpStatus:CreateWorkspaceBundleResponse', createWorkspaceBundleResponse_httpStatus - The response's http status code.

Response Lenses

createWorkspaceBundleResponse_workspaceBundle :: Lens' CreateWorkspaceBundleResponse (Maybe WorkspaceBundle)

Undocumented member.

createWorkspaceBundleResponse_httpStatus :: Lens' CreateWorkspaceBundleResponse Int

The response's http status code.