Amazonka.WorkSpaces.DeregisterWorkspaceDirectory

Description

Deregisters the specified directory. This operation is asynchronous and returns before the WorkSpace directory is deregistered. If any WorkSpaces are registered to this directory, you must remove them before you can deregister the directory.

Simple AD and AD Connector are made available to you free of charge to use with WorkSpaces. If there are no WorkSpaces being used with your Simple AD or AD Connector directory for 30 consecutive days, this directory will be automatically deregistered for use with Amazon WorkSpaces, and you will be charged for this directory as per the Directory Service pricing terms.

To delete empty directories, see Delete the Directory for Your WorkSpaces. If you delete your Simple AD or AD Connector directory, you can always create a new one when you want to start using WorkSpaces again.

Creating a Request

data DeregisterWorkspaceDirectory

See: newDeregisterWorkspaceDirectory smart constructor.

Constructors

DeregisterWorkspaceDirectory'

Fields

directoryId :: Text

The identifier of the directory. If any WorkSpaces are registered to this directory, you must remove them before you deregister the directory, or you will receive an OperationNotSupportedException error.

newDeregisterWorkspaceDirectory

Arguments

:: Text
$sel:directoryId:DeregisterWorkspaceDirectory'

-> DeregisterWorkspaceDirectory

Create a value of DeregisterWorkspaceDirectory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:DeregisterWorkspaceDirectory', deregisterWorkspaceDirectory_directoryId - The identifier of the directory. If any WorkSpaces are registered to this directory, you must remove them before you deregister the directory, or you will receive an OperationNotSupportedException error.

Request Lenses

deregisterWorkspaceDirectory_directoryId :: Lens' DeregisterWorkspaceDirectory Text

The identifier of the directory. If any WorkSpaces are registered to this directory, you must remove them before you deregister the directory, or you will receive an OperationNotSupportedException error.

Destructuring the Response

data DeregisterWorkspaceDirectoryResponse

See: newDeregisterWorkspaceDirectoryResponse smart constructor.

Constructors

DeregisterWorkspaceDirectoryResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeregisterWorkspaceDirectoryResponse

Arguments

:: Int
$sel:httpStatus:DeregisterWorkspaceDirectoryResponse'

-> DeregisterWorkspaceDirectoryResponse

Create a value of DeregisterWorkspaceDirectoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeregisterWorkspaceDirectoryResponse', deregisterWorkspaceDirectoryResponse_httpStatus - The response's http status code.

Response Lenses

deregisterWorkspaceDirectoryResponse_httpStatus :: Lens' DeregisterWorkspaceDirectoryResponse Int

The response's http status code.