Amazonka.WorkSpaces.ModifyWorkspaceState

Description

Sets the state of the specified WorkSpace.

To maintain a WorkSpace without being interrupted, set the WorkSpace state to ADMIN_MAINTENANCE. WorkSpaces in this state do not respond to requests to reboot, stop, start, rebuild, or restore. An AutoStop WorkSpace in this state is not stopped. Users cannot log into a WorkSpace in the ADMIN_MAINTENANCE state.

Creating a Request

data ModifyWorkspaceState

See: newModifyWorkspaceState smart constructor.

Constructors

ModifyWorkspaceState'

Fields

workspaceId :: Text

The identifier of the WorkSpace.

workspaceState :: TargetWorkspaceState

The WorkSpace state.

newModifyWorkspaceState

Arguments

:: Text
$sel:workspaceId:ModifyWorkspaceState'

-> TargetWorkspaceState
$sel:workspaceState:ModifyWorkspaceState'

-> ModifyWorkspaceState

Create a value of ModifyWorkspaceState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceId:ModifyWorkspaceState', modifyWorkspaceState_workspaceId - The identifier of the WorkSpace.

$sel:workspaceState:ModifyWorkspaceState', modifyWorkspaceState_workspaceState - The WorkSpace state.

Request Lenses

modifyWorkspaceState_workspaceId :: Lens' ModifyWorkspaceState Text

The identifier of the WorkSpace.

modifyWorkspaceState_workspaceState :: Lens' ModifyWorkspaceState TargetWorkspaceState

The WorkSpace state.

Destructuring the Response

data ModifyWorkspaceStateResponse

See: newModifyWorkspaceStateResponse smart constructor.

Constructors

ModifyWorkspaceStateResponse'

Fields

httpStatus :: Int

The response's http status code.

newModifyWorkspaceStateResponse

Arguments

:: Int
$sel:httpStatus:ModifyWorkspaceStateResponse'

-> ModifyWorkspaceStateResponse

Create a value of ModifyWorkspaceStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ModifyWorkspaceStateResponse', modifyWorkspaceStateResponse_httpStatus - The response's http status code.

Response Lenses

modifyWorkspaceStateResponse_httpStatus :: Lens' ModifyWorkspaceStateResponse Int

The response's http status code.