Amazonka.WorkSpaces.Types.WorkspaceAccessProperties

Description

Documentation

data WorkspaceAccessProperties

The device types and operating systems that can be used to access a WorkSpace. For more information, see Amazon WorkSpaces Client Network Requirements.

See: newWorkspaceAccessProperties smart constructor.

Constructors

WorkspaceAccessProperties'

Fields

deviceTypeWindows :: Maybe AccessPropertyValue

Indicates whether users can use Windows clients to access their WorkSpaces.

deviceTypeWeb :: Maybe AccessPropertyValue

Indicates whether users can access their WorkSpaces through a web browser.

deviceTypeAndroid :: Maybe AccessPropertyValue

Indicates whether users can use Android and Android-compatible Chrome OS devices to access their WorkSpaces.

deviceTypeLinux :: Maybe AccessPropertyValue

Indicates whether users can use Linux clients to access their WorkSpaces.

deviceTypeOsx :: Maybe AccessPropertyValue

Indicates whether users can use macOS clients to access their WorkSpaces.

deviceTypeChromeOs :: Maybe AccessPropertyValue

Indicates whether users can use Chromebooks to access their WorkSpaces.

deviceTypeIos :: Maybe AccessPropertyValue

Indicates whether users can use iOS devices to access their WorkSpaces.

deviceTypeZeroClient :: Maybe AccessPropertyValue

Indicates whether users can use zero client devices to access their WorkSpaces.

newWorkspaceAccessProperties :: WorkspaceAccessProperties

Create a value of WorkspaceAccessProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceTypeWindows:WorkspaceAccessProperties', workspaceAccessProperties_deviceTypeWindows - Indicates whether users can use Windows clients to access their WorkSpaces.

$sel:deviceTypeWeb:WorkspaceAccessProperties', workspaceAccessProperties_deviceTypeWeb - Indicates whether users can access their WorkSpaces through a web browser.

$sel:deviceTypeAndroid:WorkspaceAccessProperties', workspaceAccessProperties_deviceTypeAndroid - Indicates whether users can use Android and Android-compatible Chrome OS devices to access their WorkSpaces.

$sel:deviceTypeLinux:WorkspaceAccessProperties', workspaceAccessProperties_deviceTypeLinux - Indicates whether users can use Linux clients to access their WorkSpaces.

$sel:deviceTypeOsx:WorkspaceAccessProperties', workspaceAccessProperties_deviceTypeOsx - Indicates whether users can use macOS clients to access their WorkSpaces.

$sel:deviceTypeChromeOs:WorkspaceAccessProperties', workspaceAccessProperties_deviceTypeChromeOs - Indicates whether users can use Chromebooks to access their WorkSpaces.

$sel:deviceTypeIos:WorkspaceAccessProperties', workspaceAccessProperties_deviceTypeIos - Indicates whether users can use iOS devices to access their WorkSpaces.

$sel:deviceTypeZeroClient:WorkspaceAccessProperties', workspaceAccessProperties_deviceTypeZeroClient - Indicates whether users can use zero client devices to access their WorkSpaces.

workspaceAccessProperties_deviceTypeWindows :: Lens' WorkspaceAccessProperties (Maybe AccessPropertyValue)

Indicates whether users can use Windows clients to access their WorkSpaces.

workspaceAccessProperties_deviceTypeWeb :: Lens' WorkspaceAccessProperties (Maybe AccessPropertyValue)

Indicates whether users can access their WorkSpaces through a web browser.

workspaceAccessProperties_deviceTypeAndroid :: Lens' WorkspaceAccessProperties (Maybe AccessPropertyValue)

Indicates whether users can use Android and Android-compatible Chrome OS devices to access their WorkSpaces.

workspaceAccessProperties_deviceTypeLinux :: Lens' WorkspaceAccessProperties (Maybe AccessPropertyValue)

Indicates whether users can use Linux clients to access their WorkSpaces.

workspaceAccessProperties_deviceTypeOsx :: Lens' WorkspaceAccessProperties (Maybe AccessPropertyValue)

Indicates whether users can use macOS clients to access their WorkSpaces.

workspaceAccessProperties_deviceTypeChromeOs :: Lens' WorkspaceAccessProperties (Maybe AccessPropertyValue)

Indicates whether users can use Chromebooks to access their WorkSpaces.

workspaceAccessProperties_deviceTypeIos :: Lens' WorkspaceAccessProperties (Maybe AccessPropertyValue)

Indicates whether users can use iOS devices to access their WorkSpaces.

workspaceAccessProperties_deviceTypeZeroClient :: Lens' WorkspaceAccessProperties (Maybe AccessPropertyValue)

Indicates whether users can use zero client devices to access their WorkSpaces.