Amazonka.WorkSpaces.Types.WorkspaceDirectory

Description

Documentation

data WorkspaceDirectory

Describes a directory that is used with Amazon WorkSpaces.

See: newWorkspaceDirectory smart constructor.

Constructors

WorkspaceDirectory'

Fields

registrationCode :: Maybe Text

The registration code for the directory. This is the code that users enter in their Amazon WorkSpaces client application to connect to the directory.

iamRoleId :: Maybe Text

The identifier of the IAM role. This is the role that allows Amazon WorkSpaces to make calls to other services, such as Amazon EC2, on your behalf.

directoryId :: Maybe Text

The directory identifier.

state :: Maybe WorkspaceDirectoryState

The state of the directory's registration with Amazon WorkSpaces. After a directory is deregistered, the DEREGISTERED state is returned very briefly before the directory metadata is cleaned up, so this state is rarely returned. To confirm that a directory is deregistered, check for the directory ID by using DescribeWorkspaceDirectories. If the directory ID isn't returned, then the directory has been successfully deregistered.

customerUserName :: Maybe Text

The user name for the service account.

subnetIds :: Maybe [Text]

The identifiers of the subnets used with the directory.

ipGroupIds :: Maybe [Text]

The identifiers of the IP access control groups associated with the directory.

alias :: Maybe Text

The directory alias.

workspaceSecurityGroupId :: Maybe Text

The identifier of the security group that is assigned to new WorkSpaces.

directoryType :: Maybe WorkspaceDirectoryType

The directory type.

tenancy :: Maybe Tenancy

Specifies whether the directory is dedicated or shared. To use Bring Your Own License (BYOL), this value must be set to DEDICATED. For more information, see Bring Your Own Windows Desktop Images.

workspaceCreationProperties :: Maybe DefaultWorkspaceCreationProperties

The default creation properties for all WorkSpaces in the directory.

dnsIpAddresses :: Maybe [Text]

The IP addresses of the DNS servers for the directory.

workspaceAccessProperties :: Maybe WorkspaceAccessProperties

The devices and operating systems that users can use to access WorkSpaces.

directoryName :: Maybe Text

The name of the directory.

selfservicePermissions :: Maybe SelfservicePermissions

The default self-service permissions for WorkSpaces in the directory.

newWorkspaceDirectory :: WorkspaceDirectory

Create a value of WorkspaceDirectory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registrationCode:WorkspaceDirectory', workspaceDirectory_registrationCode - The registration code for the directory. This is the code that users enter in their Amazon WorkSpaces client application to connect to the directory.

$sel:iamRoleId:WorkspaceDirectory', workspaceDirectory_iamRoleId - The identifier of the IAM role. This is the role that allows Amazon WorkSpaces to make calls to other services, such as Amazon EC2, on your behalf.

$sel:directoryId:WorkspaceDirectory', workspaceDirectory_directoryId - The directory identifier.

$sel:state:WorkspaceDirectory', workspaceDirectory_state - The state of the directory's registration with Amazon WorkSpaces. After a directory is deregistered, the DEREGISTERED state is returned very briefly before the directory metadata is cleaned up, so this state is rarely returned. To confirm that a directory is deregistered, check for the directory ID by using DescribeWorkspaceDirectories. If the directory ID isn't returned, then the directory has been successfully deregistered.

$sel:customerUserName:WorkspaceDirectory', workspaceDirectory_customerUserName - The user name for the service account.

$sel:subnetIds:WorkspaceDirectory', workspaceDirectory_subnetIds - The identifiers of the subnets used with the directory.

$sel:ipGroupIds:WorkspaceDirectory', workspaceDirectory_ipGroupIds - The identifiers of the IP access control groups associated with the directory.

$sel:alias:WorkspaceDirectory', workspaceDirectory_alias - The directory alias.

$sel:workspaceSecurityGroupId:WorkspaceDirectory', workspaceDirectory_workspaceSecurityGroupId - The identifier of the security group that is assigned to new WorkSpaces.

$sel:directoryType:WorkspaceDirectory', workspaceDirectory_directoryType - The directory type.

$sel:tenancy:WorkspaceDirectory', workspaceDirectory_tenancy - Specifies whether the directory is dedicated or shared. To use Bring Your Own License (BYOL), this value must be set to DEDICATED. For more information, see Bring Your Own Windows Desktop Images.

$sel:workspaceCreationProperties:WorkspaceDirectory', workspaceDirectory_workspaceCreationProperties - The default creation properties for all WorkSpaces in the directory.

$sel:dnsIpAddresses:WorkspaceDirectory', workspaceDirectory_dnsIpAddresses - The IP addresses of the DNS servers for the directory.

$sel:workspaceAccessProperties:WorkspaceDirectory', workspaceDirectory_workspaceAccessProperties - The devices and operating systems that users can use to access WorkSpaces.

$sel:directoryName:WorkspaceDirectory', workspaceDirectory_directoryName - The name of the directory.

$sel:selfservicePermissions:WorkspaceDirectory', workspaceDirectory_selfservicePermissions - The default self-service permissions for WorkSpaces in the directory.

workspaceDirectory_registrationCode :: Lens' WorkspaceDirectory (Maybe Text)

The registration code for the directory. This is the code that users enter in their Amazon WorkSpaces client application to connect to the directory.

workspaceDirectory_iamRoleId :: Lens' WorkspaceDirectory (Maybe Text)

The identifier of the IAM role. This is the role that allows Amazon WorkSpaces to make calls to other services, such as Amazon EC2, on your behalf.

workspaceDirectory_directoryId :: Lens' WorkspaceDirectory (Maybe Text)

The directory identifier.

workspaceDirectory_state :: Lens' WorkspaceDirectory (Maybe WorkspaceDirectoryState)

The state of the directory's registration with Amazon WorkSpaces. After a directory is deregistered, the DEREGISTERED state is returned very briefly before the directory metadata is cleaned up, so this state is rarely returned. To confirm that a directory is deregistered, check for the directory ID by using DescribeWorkspaceDirectories. If the directory ID isn't returned, then the directory has been successfully deregistered.

workspaceDirectory_customerUserName :: Lens' WorkspaceDirectory (Maybe Text)

The user name for the service account.

workspaceDirectory_subnetIds :: Lens' WorkspaceDirectory (Maybe [Text])

The identifiers of the subnets used with the directory.

workspaceDirectory_ipGroupIds :: Lens' WorkspaceDirectory (Maybe [Text])

The identifiers of the IP access control groups associated with the directory.

workspaceDirectory_alias :: Lens' WorkspaceDirectory (Maybe Text)

The directory alias.

workspaceDirectory_workspaceSecurityGroupId :: Lens' WorkspaceDirectory (Maybe Text)

The identifier of the security group that is assigned to new WorkSpaces.

workspaceDirectory_directoryType :: Lens' WorkspaceDirectory (Maybe WorkspaceDirectoryType)

The directory type.

workspaceDirectory_tenancy :: Lens' WorkspaceDirectory (Maybe Tenancy)

Specifies whether the directory is dedicated or shared. To use Bring Your Own License (BYOL), this value must be set to DEDICATED. For more information, see Bring Your Own Windows Desktop Images.

workspaceDirectory_workspaceCreationProperties :: Lens' WorkspaceDirectory (Maybe DefaultWorkspaceCreationProperties)

The default creation properties for all WorkSpaces in the directory.

workspaceDirectory_dnsIpAddresses :: Lens' WorkspaceDirectory (Maybe [Text])

The IP addresses of the DNS servers for the directory.

workspaceDirectory_workspaceAccessProperties :: Lens' WorkspaceDirectory (Maybe WorkspaceAccessProperties)

The devices and operating systems that users can use to access WorkSpaces.

workspaceDirectory_directoryName :: Lens' WorkspaceDirectory (Maybe Text)

The name of the directory.

workspaceDirectory_selfservicePermissions :: Lens' WorkspaceDirectory (Maybe SelfservicePermissions)

The default self-service permissions for WorkSpaces in the directory.