Amazonka.XRay.GetInsight

Description

Retrieves the summary information of an insight. This includes impact to clients and root cause services, the top anomalous services, the category, the state of the insight, and the start and end time of the insight.

Creating a Request

data GetInsight

See: newGetInsight smart constructor.

Constructors

GetInsight'

Fields

insightId :: Text

The insight's unique identifier. Use the GetInsightSummaries action to retrieve an InsightId.

newGetInsight

Arguments

:: Text
$sel:insightId:GetInsight'

-> GetInsight

Create a value of GetInsight with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:insightId:GetInsight', getInsight_insightId - The insight's unique identifier. Use the GetInsightSummaries action to retrieve an InsightId.

Request Lenses

getInsight_insightId :: Lens' GetInsight Text

The insight's unique identifier. Use the GetInsightSummaries action to retrieve an InsightId.

Destructuring the Response

data GetInsightResponse

See: newGetInsightResponse smart constructor.

Constructors

GetInsightResponse'

Fields

insight :: Maybe Insight

The summary information of an insight.

httpStatus :: Int

The response's http status code.

newGetInsightResponse

Arguments

:: Int
$sel:httpStatus:GetInsightResponse'

-> GetInsightResponse

Create a value of GetInsightResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:insight:GetInsightResponse', getInsightResponse_insight - The summary information of an insight.

$sel:httpStatus:GetInsightResponse', getInsightResponse_httpStatus - The response's http status code.

Response Lenses

getInsightResponse_insight :: Lens' GetInsightResponse (Maybe Insight)

The summary information of an insight.

getInsightResponse_httpStatus :: Lens' GetInsightResponse Int

The response's http status code.