Amazonka.XRay.GetInsightEvents

Description

X-Ray reevaluates insights periodically until they're resolved, and records each intermediate state as an event. You can review an insight's events in the Impact Timeline on the Inspect page in the X-Ray console.

Creating a Request

data GetInsightEvents

See: newGetInsightEvents smart constructor.

Constructors

GetInsightEvents'

Fields

nextToken :: Maybe Text

Specify the pagination token returned by a previous request to retrieve the next page of events.

maxResults :: Maybe Natural

Used to retrieve at most the specified value of events.

insightId :: Text

The insight's unique identifier. Use the GetInsightSummaries action to retrieve an InsightId.

newGetInsightEvents

Arguments

:: Text
$sel:insightId:GetInsightEvents'

-> GetInsightEvents

Create a value of GetInsightEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetInsightEvents', getInsightEvents_nextToken - Specify the pagination token returned by a previous request to retrieve the next page of events.

$sel:maxResults:GetInsightEvents', getInsightEvents_maxResults - Used to retrieve at most the specified value of events.

$sel:insightId:GetInsightEvents', getInsightEvents_insightId - The insight's unique identifier. Use the GetInsightSummaries action to retrieve an InsightId.

Request Lenses

getInsightEvents_nextToken :: Lens' GetInsightEvents (Maybe Text)

Specify the pagination token returned by a previous request to retrieve the next page of events.

getInsightEvents_maxResults :: Lens' GetInsightEvents (Maybe Natural)

Used to retrieve at most the specified value of events.

getInsightEvents_insightId :: Lens' GetInsightEvents Text

The insight's unique identifier. Use the GetInsightSummaries action to retrieve an InsightId.

Destructuring the Response

data GetInsightEventsResponse

See: newGetInsightEventsResponse smart constructor.

Constructors

GetInsightEventsResponse'

Fields

insightEvents :: Maybe [InsightEvent]

A detailed description of the event. This includes the time of the event, client and root cause impact statistics, and the top anomalous service at the time of the event.

nextToken :: Maybe Text

Use this token to retrieve the next page of insight events.

httpStatus :: Int

The response's http status code.

newGetInsightEventsResponse

Arguments

:: Int
$sel:httpStatus:GetInsightEventsResponse'

-> GetInsightEventsResponse

Create a value of GetInsightEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:insightEvents:GetInsightEventsResponse', getInsightEventsResponse_insightEvents - A detailed description of the event. This includes the time of the event, client and root cause impact statistics, and the top anomalous service at the time of the event.

$sel:nextToken:GetInsightEvents', getInsightEventsResponse_nextToken - Use this token to retrieve the next page of insight events.

$sel:httpStatus:GetInsightEventsResponse', getInsightEventsResponse_httpStatus - The response's http status code.

Response Lenses

getInsightEventsResponse_insightEvents :: Lens' GetInsightEventsResponse (Maybe [InsightEvent])

A detailed description of the event. This includes the time of the event, client and root cause impact statistics, and the top anomalous service at the time of the event.

getInsightEventsResponse_nextToken :: Lens' GetInsightEventsResponse (Maybe Text)

Use this token to retrieve the next page of insight events.

getInsightEventsResponse_httpStatus :: Lens' GetInsightEventsResponse Int

The response's http status code.