Amazonka.XRay.GetTraceSummaries

Description

Retrieves IDs and annotations for traces available for a specified time frame using an optional filter. To get the full traces, pass the trace IDs to BatchGetTraces.

A filter expression can target traced requests that hit specific service nodes or edges, have errors, or come from a known user. For example, the following filter expression targets traces that pass through api.example.com:

service("api.example.com")

This filter expression finds traces that have an annotation named account with the value 12345:

annotation.account = "12345"

For a full list of indexed fields and keywords that you can use in filter expressions, see Using Filter Expressions in the Amazon Web Services X-Ray Developer Guide.

This operation returns paginated results.

Creating a Request

data GetTraceSummaries

See: newGetTraceSummaries smart constructor.

Constructors

GetTraceSummaries'

Fields

filterExpression :: Maybe Text

Specify a filter expression to retrieve trace summaries for services or requests that meet certain requirements.

nextToken :: Maybe Text

Specify the pagination token returned by a previous request to retrieve the next page of results.

timeRangeType :: Maybe TimeRangeType

A parameter to indicate whether to query trace summaries by TraceId or Event time.

samplingStrategy :: Maybe SamplingStrategy

A parameter to indicate whether to enable sampling on trace summaries. Input parameters are Name and Value.

sampling :: Maybe Bool

Set to true to get summaries for only a subset of available traces.

startTime :: POSIX

The start of the time frame for which to retrieve traces.

endTime :: POSIX

The end of the time frame for which to retrieve traces.

newGetTraceSummaries

Arguments

:: UTCTime
$sel:startTime:GetTraceSummaries'

-> UTCTime
$sel:endTime:GetTraceSummaries'

-> GetTraceSummaries

Create a value of GetTraceSummaries with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filterExpression:GetTraceSummaries', getTraceSummaries_filterExpression - Specify a filter expression to retrieve trace summaries for services or requests that meet certain requirements.

$sel:nextToken:GetTraceSummaries', getTraceSummaries_nextToken - Specify the pagination token returned by a previous request to retrieve the next page of results.

$sel:timeRangeType:GetTraceSummaries', getTraceSummaries_timeRangeType - A parameter to indicate whether to query trace summaries by TraceId or Event time.

$sel:samplingStrategy:GetTraceSummaries', getTraceSummaries_samplingStrategy - A parameter to indicate whether to enable sampling on trace summaries. Input parameters are Name and Value.

$sel:sampling:GetTraceSummaries', getTraceSummaries_sampling - Set to true to get summaries for only a subset of available traces.

$sel:startTime:GetTraceSummaries', getTraceSummaries_startTime - The start of the time frame for which to retrieve traces.

$sel:endTime:GetTraceSummaries', getTraceSummaries_endTime - The end of the time frame for which to retrieve traces.

Request Lenses

getTraceSummaries_filterExpression :: Lens' GetTraceSummaries (Maybe Text)

Specify a filter expression to retrieve trace summaries for services or requests that meet certain requirements.

getTraceSummaries_nextToken :: Lens' GetTraceSummaries (Maybe Text)

Specify the pagination token returned by a previous request to retrieve the next page of results.

getTraceSummaries_timeRangeType :: Lens' GetTraceSummaries (Maybe TimeRangeType)

A parameter to indicate whether to query trace summaries by TraceId or Event time.

getTraceSummaries_samplingStrategy :: Lens' GetTraceSummaries (Maybe SamplingStrategy)

A parameter to indicate whether to enable sampling on trace summaries. Input parameters are Name and Value.

getTraceSummaries_sampling :: Lens' GetTraceSummaries (Maybe Bool)

Set to true to get summaries for only a subset of available traces.

getTraceSummaries_startTime :: Lens' GetTraceSummaries UTCTime

The start of the time frame for which to retrieve traces.

getTraceSummaries_endTime :: Lens' GetTraceSummaries UTCTime

The end of the time frame for which to retrieve traces.

Destructuring the Response

data GetTraceSummariesResponse

See: newGetTraceSummariesResponse smart constructor.

Constructors

GetTraceSummariesResponse'

Fields

tracesProcessedCount :: Maybe Integer

The total number of traces processed, including traces that did not match the specified filter expression.

nextToken :: Maybe Text

If the requested time frame contained more than one page of results, you can use this token to retrieve the next page. The first page contains the most recent results, closest to the end of the time frame.

approximateTime :: Maybe POSIX

The start time of this page of results.

traceSummaries :: Maybe [TraceSummary]

Trace IDs and annotations for traces that were found in the specified time frame.

httpStatus :: Int

The response's http status code.

newGetTraceSummariesResponse

Arguments

:: Int
$sel:httpStatus:GetTraceSummariesResponse'

-> GetTraceSummariesResponse

Create a value of GetTraceSummariesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tracesProcessedCount:GetTraceSummariesResponse', getTraceSummariesResponse_tracesProcessedCount - The total number of traces processed, including traces that did not match the specified filter expression.

$sel:nextToken:GetTraceSummaries', getTraceSummariesResponse_nextToken - If the requested time frame contained more than one page of results, you can use this token to retrieve the next page. The first page contains the most recent results, closest to the end of the time frame.

$sel:approximateTime:GetTraceSummariesResponse', getTraceSummariesResponse_approximateTime - The start time of this page of results.

$sel:traceSummaries:GetTraceSummariesResponse', getTraceSummariesResponse_traceSummaries - Trace IDs and annotations for traces that were found in the specified time frame.

$sel:httpStatus:GetTraceSummariesResponse', getTraceSummariesResponse_httpStatus - The response's http status code.

Response Lenses

getTraceSummariesResponse_tracesProcessedCount :: Lens' GetTraceSummariesResponse (Maybe Integer)

The total number of traces processed, including traces that did not match the specified filter expression.

getTraceSummariesResponse_nextToken :: Lens' GetTraceSummariesResponse (Maybe Text)

If the requested time frame contained more than one page of results, you can use this token to retrieve the next page. The first page contains the most recent results, closest to the end of the time frame.

getTraceSummariesResponse_approximateTime :: Lens' GetTraceSummariesResponse (Maybe UTCTime)

The start time of this page of results.

getTraceSummariesResponse_traceSummaries :: Lens' GetTraceSummariesResponse (Maybe [TraceSummary])

Trace IDs and annotations for traces that were found in the specified time frame.

getTraceSummariesResponse_httpStatus :: Lens' GetTraceSummariesResponse Int

The response's http status code.