Amazonka.XRay.Lens

Description

Operations

PutEncryptionConfig

putEncryptionConfig_keyId :: Lens' PutEncryptionConfig (Maybe Text)

An Amazon Web Services KMS key in one of the following formats:

Alias - The name of the key. For example, alias/MyKey.

Key ID - The KMS key ID of the key. For example, ae4aa6d49-a4d8-9df9-a475-4ff6d7898456. Amazon Web Services X-Ray does not support asymmetric KMS keys.

ARN - The full Amazon Resource Name of the key ID or alias. For example, arn:aws:kms:us-east-2:123456789012:key/ae4aa6d49-a4d8-9df9-a475-4ff6d7898456. Use this format to specify a key in a different account.

Omit this key if you set Type to NONE.

putEncryptionConfig_type :: Lens' PutEncryptionConfig EncryptionType

The type of encryption. Set to KMS to use your own key for encryption. Set to NONE for default encryption.

putEncryptionConfigResponse_encryptionConfig :: Lens' PutEncryptionConfigResponse (Maybe EncryptionConfig)

The new encryption configuration.

putEncryptionConfigResponse_httpStatus :: Lens' PutEncryptionConfigResponse Int

The response's http status code.

GetServiceGraph

getServiceGraph_nextToken :: Lens' GetServiceGraph (Maybe Text)

Pagination token.

getServiceGraph_groupARN :: Lens' GetServiceGraph (Maybe Text)

The Amazon Resource Name (ARN) of a group based on which you want to generate a graph.

getServiceGraph_groupName :: Lens' GetServiceGraph (Maybe Text)

The name of a group based on which you want to generate a graph.

getServiceGraph_startTime :: Lens' GetServiceGraph UTCTime

The start of the time frame for which to generate a graph.

getServiceGraph_endTime :: Lens' GetServiceGraph UTCTime

The end of the timeframe for which to generate a graph.

getServiceGraphResponse_containsOldGroupVersions :: Lens' GetServiceGraphResponse (Maybe Bool)

A flag indicating whether the group's filter expression has been consistent, or if the returned service graph may show traces from an older version of the group's filter expression.

getServiceGraphResponse_startTime :: Lens' GetServiceGraphResponse (Maybe UTCTime)

The start of the time frame for which the graph was generated.

getServiceGraphResponse_nextToken :: Lens' GetServiceGraphResponse (Maybe Text)

Pagination token.

getServiceGraphResponse_endTime :: Lens' GetServiceGraphResponse (Maybe UTCTime)

The end of the time frame for which the graph was generated.

getServiceGraphResponse_services :: Lens' GetServiceGraphResponse (Maybe [ServiceInfo])

The services that have processed a traced request during the specified time frame.

getServiceGraphResponse_httpStatus :: Lens' GetServiceGraphResponse Int

The response's http status code.

GetSamplingTargets

getSamplingTargets_samplingStatisticsDocuments :: Lens' GetSamplingTargets [SamplingStatisticsDocument]

Information about rules that the service is using to sample requests.

getSamplingTargetsResponse_unprocessedStatistics :: Lens' GetSamplingTargetsResponse (Maybe [UnprocessedStatistics])

Information about SamplingStatisticsDocument that X-Ray could not process.

getSamplingTargetsResponse_lastRuleModification :: Lens' GetSamplingTargetsResponse (Maybe UTCTime)

The last time a user changed the sampling rule configuration. If the sampling rule configuration changed since the service last retrieved it, the service should call GetSamplingRules to get the latest version.

getSamplingTargetsResponse_samplingTargetDocuments :: Lens' GetSamplingTargetsResponse (Maybe [SamplingTargetDocument])

Updated rules that the service should use to sample requests.

getSamplingTargetsResponse_httpStatus :: Lens' GetSamplingTargetsResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_nextToken :: Lens' ListTagsForResource (Maybe Text)

A pagination token. If multiple pages of results are returned, use the NextToken value returned with the current page of results as the value of this parameter to get the next page of results.

listTagsForResource_resourceARN :: Lens' ListTagsForResource Text

The Amazon Resource Number (ARN) of an X-Ray group or sampling rule.

listTagsForResourceResponse_nextToken :: Lens' ListTagsForResourceResponse (Maybe Text)

A pagination token. If multiple pages of results are returned, use the NextToken value returned with the current page of results to get the next page of results.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

A list of tags, as key and value pairs, that is associated with the specified X-Ray group or sampling rule.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

GetTraceSummaries

getTraceSummaries_filterExpression :: Lens' GetTraceSummaries (Maybe Text)

Specify a filter expression to retrieve trace summaries for services or requests that meet certain requirements.

getTraceSummaries_nextToken :: Lens' GetTraceSummaries (Maybe Text)

Specify the pagination token returned by a previous request to retrieve the next page of results.

getTraceSummaries_timeRangeType :: Lens' GetTraceSummaries (Maybe TimeRangeType)

A parameter to indicate whether to query trace summaries by TraceId or Event time.

getTraceSummaries_samplingStrategy :: Lens' GetTraceSummaries (Maybe SamplingStrategy)

A parameter to indicate whether to enable sampling on trace summaries. Input parameters are Name and Value.

getTraceSummaries_sampling :: Lens' GetTraceSummaries (Maybe Bool)

Set to true to get summaries for only a subset of available traces.

getTraceSummaries_startTime :: Lens' GetTraceSummaries UTCTime

The start of the time frame for which to retrieve traces.

getTraceSummaries_endTime :: Lens' GetTraceSummaries UTCTime

The end of the time frame for which to retrieve traces.

getTraceSummariesResponse_tracesProcessedCount :: Lens' GetTraceSummariesResponse (Maybe Integer)

The total number of traces processed, including traces that did not match the specified filter expression.

getTraceSummariesResponse_nextToken :: Lens' GetTraceSummariesResponse (Maybe Text)

If the requested time frame contained more than one page of results, you can use this token to retrieve the next page. The first page contains the most recent results, closest to the end of the time frame.

getTraceSummariesResponse_approximateTime :: Lens' GetTraceSummariesResponse (Maybe UTCTime)

The start time of this page of results.

getTraceSummariesResponse_traceSummaries :: Lens' GetTraceSummariesResponse (Maybe [TraceSummary])

Trace IDs and annotations for traces that were found in the specified time frame.

getTraceSummariesResponse_httpStatus :: Lens' GetTraceSummariesResponse Int

The response's http status code.

PutTraceSegments

putTraceSegments_traceSegmentDocuments :: Lens' PutTraceSegments [Text]

A string containing a JSON document defining one or more segments or subsegments.

putTraceSegmentsResponse_unprocessedTraceSegments :: Lens' PutTraceSegmentsResponse (Maybe [UnprocessedTraceSegment])

Segments that failed processing.

putTraceSegmentsResponse_httpStatus :: Lens' PutTraceSegmentsResponse Int

The response's http status code.

BatchGetTraces

batchGetTraces_nextToken :: Lens' BatchGetTraces (Maybe Text)

Pagination token.

batchGetTraces_traceIds :: Lens' BatchGetTraces [Text]

Specify the trace IDs of requests for which to retrieve segments.

batchGetTracesResponse_nextToken :: Lens' BatchGetTracesResponse (Maybe Text)

Pagination token.

batchGetTracesResponse_traces :: Lens' BatchGetTracesResponse (Maybe [Trace])

Full traces for the specified requests.

batchGetTracesResponse_unprocessedTraceIds :: Lens' BatchGetTracesResponse (Maybe [Text])

Trace IDs of requests that haven't been processed.

batchGetTracesResponse_httpStatus :: Lens' BatchGetTracesResponse Int

The response's http status code.

GetInsight

getInsight_insightId :: Lens' GetInsight Text

The insight's unique identifier. Use the GetInsightSummaries action to retrieve an InsightId.

getInsightResponse_insight :: Lens' GetInsightResponse (Maybe Insight)

The summary information of an insight.

getInsightResponse_httpStatus :: Lens' GetInsightResponse Int

The response's http status code.

GetTimeSeriesServiceStatistics

getTimeSeriesServiceStatistics_entitySelectorExpression :: Lens' GetTimeSeriesServiceStatistics (Maybe Text)

A filter expression defining entities that will be aggregated for statistics. Supports ID, service, and edge functions. If no selector expression is specified, edge statistics are returned.

getTimeSeriesServiceStatistics_period :: Lens' GetTimeSeriesServiceStatistics (Maybe Int)

Aggregation period in seconds.

getTimeSeriesServiceStatistics_forecastStatistics :: Lens' GetTimeSeriesServiceStatistics (Maybe Bool)

The forecasted high and low fault count values. Forecast enabled requests require the EntitySelectorExpression ID be provided.

getTimeSeriesServiceStatistics_nextToken :: Lens' GetTimeSeriesServiceStatistics (Maybe Text)

Pagination token.

getTimeSeriesServiceStatistics_groupARN :: Lens' GetTimeSeriesServiceStatistics (Maybe Text)

The Amazon Resource Name (ARN) of the group for which to pull statistics from.

getTimeSeriesServiceStatistics_groupName :: Lens' GetTimeSeriesServiceStatistics (Maybe Text)

The case-sensitive name of the group for which to pull statistics from.

getTimeSeriesServiceStatistics_startTime :: Lens' GetTimeSeriesServiceStatistics UTCTime

The start of the time frame for which to aggregate statistics.

getTimeSeriesServiceStatistics_endTime :: Lens' GetTimeSeriesServiceStatistics UTCTime

The end of the time frame for which to aggregate statistics.

getTimeSeriesServiceStatisticsResponse_containsOldGroupVersions :: Lens' GetTimeSeriesServiceStatisticsResponse (Maybe Bool)

A flag indicating whether or not a group's filter expression has been consistent, or if a returned aggregation might show statistics from an older version of the group's filter expression.

getTimeSeriesServiceStatisticsResponse_timeSeriesServiceStatistics :: Lens' GetTimeSeriesServiceStatisticsResponse (Maybe [TimeSeriesServiceStatistics])

The collection of statistics.

getTimeSeriesServiceStatisticsResponse_nextToken :: Lens' GetTimeSeriesServiceStatisticsResponse (Maybe Text)

Pagination token.

getTimeSeriesServiceStatisticsResponse_httpStatus :: Lens' GetTimeSeriesServiceStatisticsResponse Int

The response's http status code.

GetEncryptionConfig

getEncryptionConfigResponse_encryptionConfig :: Lens' GetEncryptionConfigResponse (Maybe EncryptionConfig)

The encryption configuration document.

getEncryptionConfigResponse_httpStatus :: Lens' GetEncryptionConfigResponse Int

The response's http status code.

GetInsightImpactGraph

getInsightImpactGraph_nextToken :: Lens' GetInsightImpactGraph (Maybe Text)

Specify the pagination token returned by a previous request to retrieve the next page of results.

getInsightImpactGraph_insightId :: Lens' GetInsightImpactGraph Text

The insight's unique identifier. Use the GetInsightSummaries action to retrieve an InsightId.

getInsightImpactGraph_startTime :: Lens' GetInsightImpactGraph UTCTime

The estimated start time of the insight, in Unix time seconds. The StartTime is inclusive of the value provided and can't be more than 30 days old.

getInsightImpactGraph_endTime :: Lens' GetInsightImpactGraph UTCTime

The estimated end time of the insight, in Unix time seconds. The EndTime is exclusive of the value provided. The time range between the start time and end time can't be more than six hours.

getInsightImpactGraphResponse_serviceGraphStartTime :: Lens' GetInsightImpactGraphResponse (Maybe UTCTime)

The time, in Unix seconds, at which the service graph started.

getInsightImpactGraphResponse_startTime :: Lens' GetInsightImpactGraphResponse (Maybe UTCTime)

The provided start time.

getInsightImpactGraphResponse_insightId :: Lens' GetInsightImpactGraphResponse (Maybe Text)

The insight's unique identifier.

getInsightImpactGraphResponse_nextToken :: Lens' GetInsightImpactGraphResponse (Maybe Text)

Pagination token.

getInsightImpactGraphResponse_endTime :: Lens' GetInsightImpactGraphResponse (Maybe UTCTime)

The provided end time.

getInsightImpactGraphResponse_serviceGraphEndTime :: Lens' GetInsightImpactGraphResponse (Maybe UTCTime)

The time, in Unix seconds, at which the service graph ended.

getInsightImpactGraphResponse_services :: Lens' GetInsightImpactGraphResponse (Maybe [InsightImpactGraphService])

The Amazon Web Services instrumented services related to the insight.

getInsightImpactGraphResponse_httpStatus :: Lens' GetInsightImpactGraphResponse Int

The response's http status code.

UpdateSamplingRule

updateSamplingRule_samplingRuleUpdate :: Lens' UpdateSamplingRule SamplingRuleUpdate

The rule and fields to change.

updateSamplingRuleResponse_samplingRuleRecord :: Lens' UpdateSamplingRuleResponse (Maybe SamplingRuleRecord)

The updated rule definition and metadata.

updateSamplingRuleResponse_httpStatus :: Lens' UpdateSamplingRuleResponse Int

The response's http status code.

DeleteSamplingRule

deleteSamplingRule_ruleName :: Lens' DeleteSamplingRule (Maybe Text)

The name of the sampling rule. Specify a rule by either name or ARN, but not both.

deleteSamplingRule_ruleARN :: Lens' DeleteSamplingRule (Maybe Text)

The ARN of the sampling rule. Specify a rule by either name or ARN, but not both.

deleteSamplingRuleResponse_samplingRuleRecord :: Lens' DeleteSamplingRuleResponse (Maybe SamplingRuleRecord)

The deleted rule definition and metadata.

deleteSamplingRuleResponse_httpStatus :: Lens' DeleteSamplingRuleResponse Int

The response's http status code.

GetInsightEvents

getInsightEvents_nextToken :: Lens' GetInsightEvents (Maybe Text)

Specify the pagination token returned by a previous request to retrieve the next page of events.

getInsightEvents_maxResults :: Lens' GetInsightEvents (Maybe Natural)

Used to retrieve at most the specified value of events.

getInsightEvents_insightId :: Lens' GetInsightEvents Text

The insight's unique identifier. Use the GetInsightSummaries action to retrieve an InsightId.

getInsightEventsResponse_insightEvents :: Lens' GetInsightEventsResponse (Maybe [InsightEvent])

A detailed description of the event. This includes the time of the event, client and root cause impact statistics, and the top anomalous service at the time of the event.

getInsightEventsResponse_nextToken :: Lens' GetInsightEventsResponse (Maybe Text)

Use this token to retrieve the next page of insight events.

getInsightEventsResponse_httpStatus :: Lens' GetInsightEventsResponse Int

The response's http status code.

GetGroups

getGroups_nextToken :: Lens' GetGroups (Maybe Text)

Pagination token.

getGroupsResponse_groups :: Lens' GetGroupsResponse (Maybe [GroupSummary])

The collection of all active groups.

getGroupsResponse_nextToken :: Lens' GetGroupsResponse (Maybe Text)

Pagination token.

getGroupsResponse_httpStatus :: Lens' GetGroupsResponse Int

The response's http status code.

GetInsightSummaries

getInsightSummaries_states :: Lens' GetInsightSummaries (Maybe [InsightState])

The list of insight states.

getInsightSummaries_nextToken :: Lens' GetInsightSummaries (Maybe Text)

Pagination token.

getInsightSummaries_groupARN :: Lens' GetInsightSummaries (Maybe Text)

The Amazon Resource Name (ARN) of the group. Required if the GroupName isn't provided.

getInsightSummaries_groupName :: Lens' GetInsightSummaries (Maybe Text)

The name of the group. Required if the GroupARN isn't provided.

getInsightSummaries_maxResults :: Lens' GetInsightSummaries (Maybe Natural)

The maximum number of results to display.

getInsightSummaries_startTime :: Lens' GetInsightSummaries UTCTime

The beginning of the time frame in which the insights started. The start time can't be more than 30 days old.

getInsightSummaries_endTime :: Lens' GetInsightSummaries UTCTime

The end of the time frame in which the insights ended. The end time can't be more than 30 days old.

getInsightSummariesResponse_insightSummaries :: Lens' GetInsightSummariesResponse (Maybe [InsightSummary])

The summary of each insight within the group matching the provided filters. The summary contains the InsightID, start and end time, the root cause service, the root cause and client impact statistics, the top anomalous services, and the status of the insight.

getInsightSummariesResponse_nextToken :: Lens' GetInsightSummariesResponse (Maybe Text)

Pagination token.

getInsightSummariesResponse_httpStatus :: Lens' GetInsightSummariesResponse Int

The response's http status code.

PutTelemetryRecords

putTelemetryRecords_hostname :: Lens' PutTelemetryRecords (Maybe Text)

putTelemetryRecords_eC2InstanceId :: Lens' PutTelemetryRecords (Maybe Text)

putTelemetryRecords_resourceARN :: Lens' PutTelemetryRecords (Maybe Text)

putTelemetryRecords_telemetryRecords :: Lens' PutTelemetryRecords [TelemetryRecord]

putTelemetryRecordsResponse_httpStatus :: Lens' PutTelemetryRecordsResponse Int

The response's http status code.

GetSamplingRules

getSamplingRules_nextToken :: Lens' GetSamplingRules (Maybe Text)

Pagination token.

getSamplingRulesResponse_samplingRuleRecords :: Lens' GetSamplingRulesResponse (Maybe [SamplingRuleRecord])

Rule definitions and metadata.

getSamplingRulesResponse_nextToken :: Lens' GetSamplingRulesResponse (Maybe Text)

Pagination token.

getSamplingRulesResponse_httpStatus :: Lens' GetSamplingRulesResponse Int

The response's http status code.

TagResource

tagResource_resourceARN :: Lens' TagResource Text

The Amazon Resource Number (ARN) of an X-Ray group or sampling rule.

tagResource_tags :: Lens' TagResource [Tag]

A map that contains one or more tag keys and tag values to attach to an X-Ray group or sampling rule. For more information about ways to use tags, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference.

The following restrictions apply to tags:

Maximum number of user-applied tags per resource: 50

Maximum tag key length: 128 Unicode characters

Maximum tag value length: 256 Unicode characters

Valid values for key and value: a-z, A-Z, 0-9, space, and the following characters: _ . : / = + - and @

Tag keys and values are case sensitive.

Don't use aws: as a prefix for keys; it's reserved for Amazon Web Services use. You cannot edit or delete system tags.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

GetTraceGraph

getTraceGraph_nextToken :: Lens' GetTraceGraph (Maybe Text)

Pagination token.

getTraceGraph_traceIds :: Lens' GetTraceGraph [Text]

Trace IDs of requests for which to generate a service graph.

getTraceGraphResponse_nextToken :: Lens' GetTraceGraphResponse (Maybe Text)

Pagination token.

getTraceGraphResponse_services :: Lens' GetTraceGraphResponse (Maybe [ServiceInfo])

The services that have processed one of the specified requests.

getTraceGraphResponse_httpStatus :: Lens' GetTraceGraphResponse Int

The response's http status code.

CreateGroup

createGroup_filterExpression :: Lens' CreateGroup (Maybe Text)

The filter expression defining criteria by which to group traces.

createGroup_insightsConfiguration :: Lens' CreateGroup (Maybe InsightsConfiguration)

The structure containing configurations related to insights.

The InsightsEnabled boolean can be set to true to enable insights for the new group or false to disable insights for the new group.

The NotifcationsEnabled boolean can be set to true to enable insights notifications for the new group. Notifications may only be enabled on a group with InsightsEnabled set to true.

createGroup_tags :: Lens' CreateGroup (Maybe [Tag])

A map that contains one or more tag keys and tag values to attach to an X-Ray group. For more information about ways to use tags, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference.

The following restrictions apply to tags:

Maximum number of user-applied tags per resource: 50

Maximum tag key length: 128 Unicode characters

Maximum tag value length: 256 Unicode characters

Valid values for key and value: a-z, A-Z, 0-9, space, and the following characters: _ . : / = + - and @

Tag keys and values are case sensitive.

Don't use aws: as a prefix for keys; it's reserved for Amazon Web Services use.

createGroup_groupName :: Lens' CreateGroup Text

The case-sensitive name of the new group. Default is a reserved name and names must be unique.

createGroupResponse_group :: Lens' CreateGroupResponse (Maybe Group)

The group that was created. Contains the name of the group that was created, the Amazon Resource Name (ARN) of the group that was generated based on the group name, the filter expression, and the insight configuration that was assigned to the group.

createGroupResponse_httpStatus :: Lens' CreateGroupResponse Int

The response's http status code.

UntagResource

untagResource_resourceARN :: Lens' UntagResource Text

The Amazon Resource Number (ARN) of an X-Ray group or sampling rule.

untagResource_tagKeys :: Lens' UntagResource [Text]

Keys for one or more tags that you want to remove from an X-Ray group or sampling rule.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DeleteGroup

deleteGroup_groupARN :: Lens' DeleteGroup (Maybe Text)

The ARN of the group that was generated on creation.

deleteGroup_groupName :: Lens' DeleteGroup (Maybe Text)

The case-sensitive name of the group.

deleteGroupResponse_httpStatus :: Lens' DeleteGroupResponse Int

The response's http status code.

UpdateGroup

updateGroup_filterExpression :: Lens' UpdateGroup (Maybe Text)

The updated filter expression defining criteria by which to group traces.

updateGroup_insightsConfiguration :: Lens' UpdateGroup (Maybe InsightsConfiguration)

The structure containing configurations related to insights.

The InsightsEnabled boolean can be set to true to enable insights for the group or false to disable insights for the group.

The NotifcationsEnabled boolean can be set to true to enable insights notifications for the group. Notifications can only be enabled on a group with InsightsEnabled set to true.

updateGroup_groupARN :: Lens' UpdateGroup (Maybe Text)

The ARN that was generated upon creation.

updateGroup_groupName :: Lens' UpdateGroup (Maybe Text)

The case-sensitive name of the group.

updateGroupResponse_group :: Lens' UpdateGroupResponse (Maybe Group)

The group that was updated. Contains the name of the group that was updated, the ARN of the group that was updated, the updated filter expression, and the updated insight configuration assigned to the group.

updateGroupResponse_httpStatus :: Lens' UpdateGroupResponse Int

The response's http status code.

GetGroup

getGroup_groupARN :: Lens' GetGroup (Maybe Text)

The ARN of the group that was generated on creation.

getGroup_groupName :: Lens' GetGroup (Maybe Text)

The case-sensitive name of the group.

getGroupResponse_group :: Lens' GetGroupResponse (Maybe Group)

The group that was requested. Contains the name of the group, the ARN of the group, the filter expression, and the insight configuration assigned to the group.

getGroupResponse_httpStatus :: Lens' GetGroupResponse Int

The response's http status code.

GetSamplingStatisticSummaries

getSamplingStatisticSummaries_nextToken :: Lens' GetSamplingStatisticSummaries (Maybe Text)

Pagination token.

getSamplingStatisticSummariesResponse_samplingStatisticSummaries :: Lens' GetSamplingStatisticSummariesResponse (Maybe [SamplingStatisticSummary])

Information about the number of requests instrumented for each sampling rule.

getSamplingStatisticSummariesResponse_nextToken :: Lens' GetSamplingStatisticSummariesResponse (Maybe Text)

Pagination token.

getSamplingStatisticSummariesResponse_httpStatus :: Lens' GetSamplingStatisticSummariesResponse Int

The response's http status code.

CreateSamplingRule

createSamplingRule_tags :: Lens' CreateSamplingRule (Maybe [Tag])

A map that contains one or more tag keys and tag values to attach to an X-Ray sampling rule. For more information about ways to use tags, see Tagging Amazon Web Services resources in the Amazon Web Services General Reference.

The following restrictions apply to tags:

Maximum number of user-applied tags per resource: 50

Maximum tag key length: 128 Unicode characters

Maximum tag value length: 256 Unicode characters

Valid values for key and value: a-z, A-Z, 0-9, space, and the following characters: _ . : / = + - and @

Tag keys and values are case sensitive.

Don't use aws: as a prefix for keys; it's reserved for Amazon Web Services use.

createSamplingRule_samplingRule :: Lens' CreateSamplingRule SamplingRule

The rule definition.

createSamplingRuleResponse_samplingRuleRecord :: Lens' CreateSamplingRuleResponse (Maybe SamplingRuleRecord)

The saved rule definition and metadata.

createSamplingRuleResponse_httpStatus :: Lens' CreateSamplingRuleResponse Int

The response's http status code.

Types

Alias

alias_names :: Lens' Alias (Maybe [Text])

A list of names for the alias, including the canonical name.

alias_name :: Lens' Alias (Maybe Text)

The canonical name of the alias.

alias_type :: Lens' Alias (Maybe Text)

The type of the alias.

AnnotationValue

annotationValue_numberValue :: Lens' AnnotationValue (Maybe Double)

Value for a Number annotation.

annotationValue_stringValue :: Lens' AnnotationValue (Maybe Text)

Value for a String annotation.

annotationValue_booleanValue :: Lens' AnnotationValue (Maybe Bool)

Value for a Boolean annotation.

AnomalousService

anomalousService_serviceId :: Lens' AnomalousService (Maybe ServiceId)

Undocumented member.

AvailabilityZoneDetail

availabilityZoneDetail_name :: Lens' AvailabilityZoneDetail (Maybe Text)

The name of a corresponding Availability Zone.

BackendConnectionErrors

backendConnectionErrors_otherCount :: Lens' BackendConnectionErrors (Maybe Int)

backendConnectionErrors_timeoutCount :: Lens' BackendConnectionErrors (Maybe Int)

backendConnectionErrors_hTTPCode5XXCount :: Lens' BackendConnectionErrors (Maybe Int)

backendConnectionErrors_connectionRefusedCount :: Lens' BackendConnectionErrors (Maybe Int)

backendConnectionErrors_hTTPCode4XXCount :: Lens' BackendConnectionErrors (Maybe Int)

backendConnectionErrors_unknownHostCount :: Lens' BackendConnectionErrors (Maybe Int)

Edge

edge_startTime :: Lens' Edge (Maybe UTCTime)

The start time of the first segment on the edge.

edge_aliases :: Lens' Edge (Maybe [Alias])

Aliases for the edge.

edge_responseTimeHistogram :: Lens' Edge (Maybe [HistogramEntry])

A histogram that maps the spread of client response times on an edge.

edge_referenceId :: Lens' Edge (Maybe Int)

Identifier of the edge. Unique within a service map.

edge_endTime :: Lens' Edge (Maybe UTCTime)

The end time of the last segment on the edge.

edge_summaryStatistics :: Lens' Edge (Maybe EdgeStatistics)

Response statistics for segments on the edge.

EdgeStatistics

edgeStatistics_faultStatistics :: Lens' EdgeStatistics (Maybe FaultStatistics)

Information about requests that failed with a 5xx Server Error status code.

edgeStatistics_okCount :: Lens' EdgeStatistics (Maybe Integer)

The number of requests that completed with a 2xx Success status code.

edgeStatistics_totalResponseTime :: Lens' EdgeStatistics (Maybe Double)

The aggregate response time of completed requests.

edgeStatistics_errorStatistics :: Lens' EdgeStatistics (Maybe ErrorStatistics)

Information about requests that failed with a 4xx Client Error status code.

edgeStatistics_totalCount :: Lens' EdgeStatistics (Maybe Integer)

The total number of completed requests.

EncryptionConfig

encryptionConfig_status :: Lens' EncryptionConfig (Maybe EncryptionStatus)

The encryption status. While the status is UPDATING, X-Ray may encrypt data with a combination of the new and old settings.

encryptionConfig_keyId :: Lens' EncryptionConfig (Maybe Text)

The ID of the KMS key used for encryption, if applicable.

encryptionConfig_type :: Lens' EncryptionConfig (Maybe EncryptionType)

The type of encryption. Set to KMS for encryption with KMS keys. Set to NONE for default encryption.

ErrorRootCause

errorRootCause_clientImpacting :: Lens' ErrorRootCause (Maybe Bool)

A flag that denotes that the root cause impacts the trace client.

errorRootCause_services :: Lens' ErrorRootCause (Maybe [ErrorRootCauseService])

A list of services corresponding to an error. A service identifies a segment and it contains a name, account ID, type, and inferred flag.

ErrorRootCauseEntity

errorRootCauseEntity_exceptions :: Lens' ErrorRootCauseEntity (Maybe [RootCauseException])

The types and messages of the exceptions.

errorRootCauseEntity_remote :: Lens' ErrorRootCauseEntity (Maybe Bool)

A flag that denotes a remote subsegment.

errorRootCauseEntity_name :: Lens' ErrorRootCauseEntity (Maybe Text)

The name of the entity.

ErrorRootCauseService

errorRootCauseService_entityPath :: Lens' ErrorRootCauseService (Maybe [ErrorRootCauseEntity])

The path of root cause entities found on the service.

errorRootCauseService_accountId :: Lens' ErrorRootCauseService (Maybe Text)

The account ID associated to the service.

errorRootCauseService_names :: Lens' ErrorRootCauseService (Maybe [Text])

A collection of associated service names.

errorRootCauseService_name :: Lens' ErrorRootCauseService (Maybe Text)

The service name.

errorRootCauseService_inferred :: Lens' ErrorRootCauseService (Maybe Bool)

A Boolean value indicating if the service is inferred from the trace.

errorRootCauseService_type :: Lens' ErrorRootCauseService (Maybe Text)

The type associated to the service.

ErrorStatistics

errorStatistics_otherCount :: Lens' ErrorStatistics (Maybe Integer)

The number of requests that failed with untracked 4xx Client Error status codes.

errorStatistics_throttleCount :: Lens' ErrorStatistics (Maybe Integer)

The number of requests that failed with a 419 throttling status code.

errorStatistics_totalCount :: Lens' ErrorStatistics (Maybe Integer)

The total number of requests that failed with a 4xx Client Error status code.

FaultRootCause

faultRootCause_clientImpacting :: Lens' FaultRootCause (Maybe Bool)

A flag that denotes that the root cause impacts the trace client.

faultRootCause_services :: Lens' FaultRootCause (Maybe [FaultRootCauseService])

A list of corresponding services. A service identifies a segment and it contains a name, account ID, type, and inferred flag.

FaultRootCauseEntity

faultRootCauseEntity_exceptions :: Lens' FaultRootCauseEntity (Maybe [RootCauseException])

The types and messages of the exceptions.

faultRootCauseEntity_remote :: Lens' FaultRootCauseEntity (Maybe Bool)

A flag that denotes a remote subsegment.

faultRootCauseEntity_name :: Lens' FaultRootCauseEntity (Maybe Text)

The name of the entity.

FaultRootCauseService

faultRootCauseService_entityPath :: Lens' FaultRootCauseService (Maybe [FaultRootCauseEntity])

The path of root cause entities found on the service.

faultRootCauseService_accountId :: Lens' FaultRootCauseService (Maybe Text)

The account ID associated to the service.

faultRootCauseService_names :: Lens' FaultRootCauseService (Maybe [Text])

A collection of associated service names.

faultRootCauseService_name :: Lens' FaultRootCauseService (Maybe Text)

The service name.

faultRootCauseService_inferred :: Lens' FaultRootCauseService (Maybe Bool)

A Boolean value indicating if the service is inferred from the trace.

faultRootCauseService_type :: Lens' FaultRootCauseService (Maybe Text)

The type associated to the service.

FaultStatistics

faultStatistics_otherCount :: Lens' FaultStatistics (Maybe Integer)

The number of requests that failed with untracked 5xx Server Error status codes.

faultStatistics_totalCount :: Lens' FaultStatistics (Maybe Integer)

The total number of requests that failed with a 5xx Server Error status code.

ForecastStatistics

forecastStatistics_faultCountLow :: Lens' ForecastStatistics (Maybe Integer)

The lower limit of fault counts for a service.

forecastStatistics_faultCountHigh :: Lens' ForecastStatistics (Maybe Integer)

The upper limit of fault counts for a service.

Group

group_filterExpression :: Lens' Group (Maybe Text)

The filter expression defining the parameters to include traces.

group_insightsConfiguration :: Lens' Group (Maybe InsightsConfiguration)

The structure containing configurations related to insights.

The InsightsEnabled boolean can be set to true to enable insights for the group or false to disable insights for the group.

The NotificationsEnabled boolean can be set to true to enable insights notifications through Amazon EventBridge for the group.

group_groupARN :: Lens' Group (Maybe Text)

The Amazon Resource Name (ARN) of the group generated based on the GroupName.

group_groupName :: Lens' Group (Maybe Text)

The unique case-sensitive name of the group.

GroupSummary

groupSummary_filterExpression :: Lens' GroupSummary (Maybe Text)

The filter expression defining the parameters to include traces.

groupSummary_insightsConfiguration :: Lens' GroupSummary (Maybe InsightsConfiguration)

The structure containing configurations related to insights.

The InsightsEnabled boolean can be set to true to enable insights for the group or false to disable insights for the group.

The NotificationsEnabled boolean can be set to true to enable insights notifications. Notifications can only be enabled on a group with InsightsEnabled set to true.

groupSummary_groupARN :: Lens' GroupSummary (Maybe Text)

The ARN of the group generated based on the GroupName.

groupSummary_groupName :: Lens' GroupSummary (Maybe Text)

The unique case-sensitive name of the group.

HistogramEntry

histogramEntry_count :: Lens' HistogramEntry (Maybe Int)

The prevalence of the entry.

histogramEntry_value :: Lens' HistogramEntry (Maybe Double)

The value of the entry.

Http

http_httpMethod :: Lens' Http (Maybe Text)

The request method.

http_httpStatus :: Lens' Http (Maybe Int)

The response status.

http_clientIp :: Lens' Http (Maybe Text)

The IP address of the requestor.

http_userAgent :: Lens' Http (Maybe Text)

The request's user agent string.

http_httpURL :: Lens' Http (Maybe Text)

The request URL.

Insight

insight_summary :: Lens' Insight (Maybe Text)

A brief description of the insight.

insight_state :: Lens' Insight (Maybe InsightState)

The current state of the insight.

insight_startTime :: Lens' Insight (Maybe UTCTime)

The time, in Unix seconds, at which the insight began.

insight_insightId :: Lens' Insight (Maybe Text)

The insights unique identifier.

insight_categories :: Lens' Insight (Maybe [InsightCategory])

The categories that label and describe the type of insight.

insight_rootCauseServiceRequestImpactStatistics :: Lens' Insight (Maybe RequestImpactStatistics)

The impact statistics of the root cause service. This includes the number of requests to the client service and whether the requests were faults or okay.

insight_topAnomalousServices :: Lens' Insight (Maybe [AnomalousService])

The service within the insight that is most impacted by the incident.

insight_rootCauseServiceId :: Lens' Insight (Maybe ServiceId)

Undocumented member.

insight_clientRequestImpactStatistics :: Lens' Insight (Maybe RequestImpactStatistics)

The impact statistics of the client side service. This includes the number of requests to the client service and whether the requests were faults or okay.

insight_endTime :: Lens' Insight (Maybe UTCTime)

The time, in Unix seconds, at which the insight ended.

insight_groupARN :: Lens' Insight (Maybe Text)

The Amazon Resource Name (ARN) of the group that the insight belongs to.

insight_groupName :: Lens' Insight (Maybe Text)

The name of the group that the insight belongs to.

InsightEvent

insightEvent_summary :: Lens' InsightEvent (Maybe Text)

A brief description of the event.

insightEvent_eventTime :: Lens' InsightEvent (Maybe UTCTime)

The time, in Unix seconds, at which the event was recorded.

insightEvent_rootCauseServiceRequestImpactStatistics :: Lens' InsightEvent (Maybe RequestImpactStatistics)

The impact statistics of the root cause service. This includes the number of requests to the client service and whether the requests were faults or okay.

insightEvent_topAnomalousServices :: Lens' InsightEvent (Maybe [AnomalousService])

The service during the event that is most impacted by the incident.

insightEvent_clientRequestImpactStatistics :: Lens' InsightEvent (Maybe RequestImpactStatistics)

The impact statistics of the client side service. This includes the number of requests to the client service and whether the requests were faults or okay.

InsightImpactGraphEdge

insightImpactGraphEdge_referenceId :: Lens' InsightImpactGraphEdge (Maybe Int)

Identifier of the edge. Unique within a service map.

InsightImpactGraphService

insightImpactGraphService_referenceId :: Lens' InsightImpactGraphService (Maybe Int)

Identifier for the service. Unique within the service map.

insightImpactGraphService_accountId :: Lens' InsightImpactGraphService (Maybe Text)

Identifier of the Amazon Web Services account in which the service runs.

insightImpactGraphService_names :: Lens' InsightImpactGraphService (Maybe [Text])

A list of names for the service, including the canonical name.

insightImpactGraphService_name :: Lens' InsightImpactGraphService (Maybe Text)

The canonical name of the service.

insightImpactGraphService_type :: Lens' InsightImpactGraphService (Maybe Text)

Identifier for the service. Unique within the service map.

Amazon Web Services Resource - The type of an Amazon Web Services resource. For example, AWS::EC2::Instance for an application running on Amazon EC2 or AWS::DynamoDB::Table for an Amazon DynamoDB table that the application used.

Amazon Web Services Service - The type of an Amazon Web Services service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table.

Amazon Web Services Service - The type of an Amazon Web Services service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table.

remote - A downstream service of indeterminate type.

insightImpactGraphService_edges :: Lens' InsightImpactGraphService (Maybe [InsightImpactGraphEdge])

Connections to downstream services.

InsightSummary

insightSummary_summary :: Lens' InsightSummary (Maybe Text)

A brief description of the insight.

insightSummary_state :: Lens' InsightSummary (Maybe InsightState)

The current state of the insight.

insightSummary_startTime :: Lens' InsightSummary (Maybe UTCTime)

The time, in Unix seconds, at which the insight began.

insightSummary_insightId :: Lens' InsightSummary (Maybe Text)

The insights unique identifier.

insightSummary_categories :: Lens' InsightSummary (Maybe [InsightCategory])

Categories The categories that label and describe the type of insight.

insightSummary_rootCauseServiceRequestImpactStatistics :: Lens' InsightSummary (Maybe RequestImpactStatistics)

The impact statistics of the root cause service. This includes the number of requests to the client service and whether the requests were faults or okay.

insightSummary_topAnomalousServices :: Lens' InsightSummary (Maybe [AnomalousService])

The service within the insight that is most impacted by the incident.

insightSummary_rootCauseServiceId :: Lens' InsightSummary (Maybe ServiceId)

Undocumented member.

insightSummary_clientRequestImpactStatistics :: Lens' InsightSummary (Maybe RequestImpactStatistics)

The impact statistics of the client side service. This includes the number of requests to the client service and whether the requests were faults or okay.

insightSummary_endTime :: Lens' InsightSummary (Maybe UTCTime)

The time, in Unix seconds, at which the insight ended.

insightSummary_groupARN :: Lens' InsightSummary (Maybe Text)

The Amazon Resource Name (ARN) of the group that the insight belongs to.

insightSummary_groupName :: Lens' InsightSummary (Maybe Text)

The name of the group that the insight belongs to.

insightSummary_lastUpdateTime :: Lens' InsightSummary (Maybe UTCTime)

The time, in Unix seconds, that the insight was last updated.

InsightsConfiguration

insightsConfiguration_notificationsEnabled :: Lens' InsightsConfiguration (Maybe Bool)

Set the NotificationsEnabled value to true to enable insights notifications. Notifications can only be enabled on a group with InsightsEnabled set to true.

insightsConfiguration_insightsEnabled :: Lens' InsightsConfiguration (Maybe Bool)

Set the InsightsEnabled value to true to enable insights or false to disable insights.

InstanceIdDetail

instanceIdDetail_id :: Lens' InstanceIdDetail (Maybe Text)

The ID of a corresponding EC2 instance.

RequestImpactStatistics

requestImpactStatistics_okCount :: Lens' RequestImpactStatistics (Maybe Integer)

The number of successful requests.

requestImpactStatistics_faultCount :: Lens' RequestImpactStatistics (Maybe Integer)

The number of requests that have resulted in a fault,

requestImpactStatistics_totalCount :: Lens' RequestImpactStatistics (Maybe Integer)

The total number of requests to the service.

ResourceARNDetail

resourceARNDetail_arn :: Lens' ResourceARNDetail (Maybe Text)

The ARN of a corresponding resource.

ResponseTimeRootCause

responseTimeRootCause_clientImpacting :: Lens' ResponseTimeRootCause (Maybe Bool)

A flag that denotes that the root cause impacts the trace client.

responseTimeRootCause_services :: Lens' ResponseTimeRootCause (Maybe [ResponseTimeRootCauseService])

A list of corresponding services. A service identifies a segment and contains a name, account ID, type, and inferred flag.

ResponseTimeRootCauseEntity

responseTimeRootCauseEntity_remote :: Lens' ResponseTimeRootCauseEntity (Maybe Bool)

A flag that denotes a remote subsegment.

responseTimeRootCauseEntity_coverage :: Lens' ResponseTimeRootCauseEntity (Maybe Double)

The type and messages of the exceptions.

responseTimeRootCauseEntity_name :: Lens' ResponseTimeRootCauseEntity (Maybe Text)

The name of the entity.

ResponseTimeRootCauseService

responseTimeRootCauseService_entityPath :: Lens' ResponseTimeRootCauseService (Maybe [ResponseTimeRootCauseEntity])

The path of root cause entities found on the service.

responseTimeRootCauseService_accountId :: Lens' ResponseTimeRootCauseService (Maybe Text)

The account ID associated to the service.

responseTimeRootCauseService_names :: Lens' ResponseTimeRootCauseService (Maybe [Text])

A collection of associated service names.

responseTimeRootCauseService_name :: Lens' ResponseTimeRootCauseService (Maybe Text)

The service name.

responseTimeRootCauseService_inferred :: Lens' ResponseTimeRootCauseService (Maybe Bool)

A Boolean value indicating if the service is inferred from the trace.

responseTimeRootCauseService_type :: Lens' ResponseTimeRootCauseService (Maybe Text)

The type associated to the service.

RootCauseException

rootCauseException_name :: Lens' RootCauseException (Maybe Text)

The name of the exception.

rootCauseException_message :: Lens' RootCauseException (Maybe Text)

The message of the exception.

SamplingRule

samplingRule_ruleName :: Lens' SamplingRule (Maybe Text)

The name of the sampling rule. Specify a rule by either name or ARN, but not both.

samplingRule_attributes :: Lens' SamplingRule (Maybe (HashMap Text Text))

Matches attributes derived from the request.

samplingRule_ruleARN :: Lens' SamplingRule (Maybe Text)

The ARN of the sampling rule. Specify a rule by either name or ARN, but not both.

samplingRule_resourceARN :: Lens' SamplingRule Text

Matches the ARN of the Amazon Web Services resource on which the service runs.

samplingRule_priority :: Lens' SamplingRule Natural

The priority of the sampling rule.

samplingRule_fixedRate :: Lens' SamplingRule Double

The percentage of matching requests to instrument, after the reservoir is exhausted.

samplingRule_reservoirSize :: Lens' SamplingRule Natural

A fixed number of matching requests to instrument per second, prior to applying the fixed rate. The reservoir is not used directly by services, but applies to all services using the rule collectively.

samplingRule_serviceName :: Lens' SamplingRule Text

Matches the name that the service uses to identify itself in segments.

samplingRule_serviceType :: Lens' SamplingRule Text

Matches the origin that the service uses to identify its type in segments.

samplingRule_host :: Lens' SamplingRule Text

Matches the hostname from a request URL.

samplingRule_hTTPMethod :: Lens' SamplingRule Text

Matches the HTTP method of a request.

samplingRule_uRLPath :: Lens' SamplingRule Text

Matches the path from a request URL.

samplingRule_version :: Lens' SamplingRule Natural

The version of the sampling rule format (1).

SamplingRuleRecord

samplingRuleRecord_modifiedAt :: Lens' SamplingRuleRecord (Maybe UTCTime)

When the rule was last modified.

samplingRuleRecord_samplingRule :: Lens' SamplingRuleRecord (Maybe SamplingRule)

The sampling rule.

samplingRuleRecord_createdAt :: Lens' SamplingRuleRecord (Maybe UTCTime)

When the rule was created.

SamplingRuleUpdate

samplingRuleUpdate_hTTPMethod :: Lens' SamplingRuleUpdate (Maybe Text)

Matches the HTTP method of a request.

samplingRuleUpdate_priority :: Lens' SamplingRuleUpdate (Maybe Int)

The priority of the sampling rule.

samplingRuleUpdate_ruleName :: Lens' SamplingRuleUpdate (Maybe Text)

The name of the sampling rule. Specify a rule by either name or ARN, but not both.

samplingRuleUpdate_reservoirSize :: Lens' SamplingRuleUpdate (Maybe Int)

A fixed number of matching requests to instrument per second, prior to applying the fixed rate. The reservoir is not used directly by services, but applies to all services using the rule collectively.

samplingRuleUpdate_fixedRate :: Lens' SamplingRuleUpdate (Maybe Double)

The percentage of matching requests to instrument, after the reservoir is exhausted.

samplingRuleUpdate_resourceARN :: Lens' SamplingRuleUpdate (Maybe Text)

Matches the ARN of the Amazon Web Services resource on which the service runs.

samplingRuleUpdate_attributes :: Lens' SamplingRuleUpdate (Maybe (HashMap Text Text))

Matches attributes derived from the request.

samplingRuleUpdate_serviceName :: Lens' SamplingRuleUpdate (Maybe Text)

Matches the name that the service uses to identify itself in segments.

samplingRuleUpdate_serviceType :: Lens' SamplingRuleUpdate (Maybe Text)

Matches the origin that the service uses to identify its type in segments.

samplingRuleUpdate_host :: Lens' SamplingRuleUpdate (Maybe Text)

Matches the hostname from a request URL.

samplingRuleUpdate_ruleARN :: Lens' SamplingRuleUpdate (Maybe Text)

The ARN of the sampling rule. Specify a rule by either name or ARN, but not both.

samplingRuleUpdate_uRLPath :: Lens' SamplingRuleUpdate (Maybe Text)

Matches the path from a request URL.

SamplingStatisticSummary

samplingStatisticSummary_requestCount :: Lens' SamplingStatisticSummary (Maybe Int)

The number of requests that matched the rule.

samplingStatisticSummary_borrowCount :: Lens' SamplingStatisticSummary (Maybe Int)

The number of requests recorded with borrowed reservoir quota.

samplingStatisticSummary_ruleName :: Lens' SamplingStatisticSummary (Maybe Text)

The name of the sampling rule.

samplingStatisticSummary_timestamp :: Lens' SamplingStatisticSummary (Maybe UTCTime)

The start time of the reporting window.

samplingStatisticSummary_sampledCount :: Lens' SamplingStatisticSummary (Maybe Int)

The number of requests recorded.

SamplingStatisticsDocument

samplingStatisticsDocument_borrowCount :: Lens' SamplingStatisticsDocument (Maybe Natural)

The number of requests recorded with borrowed reservoir quota.

samplingStatisticsDocument_ruleName :: Lens' SamplingStatisticsDocument Text

The name of the sampling rule.

samplingStatisticsDocument_clientID :: Lens' SamplingStatisticsDocument Text

A unique identifier for the service in hexadecimal.

samplingStatisticsDocument_timestamp :: Lens' SamplingStatisticsDocument UTCTime

The current time.

samplingStatisticsDocument_requestCount :: Lens' SamplingStatisticsDocument Natural

The number of requests that matched the rule.

samplingStatisticsDocument_sampledCount :: Lens' SamplingStatisticsDocument Natural

The number of requests recorded.

SamplingStrategy

samplingStrategy_value :: Lens' SamplingStrategy (Maybe Double)

The value of a sampling rule.

samplingStrategy_name :: Lens' SamplingStrategy (Maybe SamplingStrategyName)

The name of a sampling rule.

SamplingTargetDocument

samplingTargetDocument_reservoirQuota :: Lens' SamplingTargetDocument (Maybe Int)

The number of requests per second that X-Ray allocated for this service.

samplingTargetDocument_ruleName :: Lens' SamplingTargetDocument (Maybe Text)

The name of the sampling rule.

samplingTargetDocument_fixedRate :: Lens' SamplingTargetDocument (Maybe Double)

The percentage of matching requests to instrument, after the reservoir is exhausted.

samplingTargetDocument_interval :: Lens' SamplingTargetDocument (Maybe Int)

The number of seconds for the service to wait before getting sampling targets again.

samplingTargetDocument_reservoirQuotaTTL :: Lens' SamplingTargetDocument (Maybe UTCTime)

When the reservoir quota expires.

Segment

segment_document :: Lens' Segment (Maybe Text)

The segment document.

segment_id :: Lens' Segment (Maybe Text)

The segment's ID.

ServiceId

serviceId_accountId :: Lens' ServiceId (Maybe Text)

serviceId_names :: Lens' ServiceId (Maybe [Text])

serviceId_name :: Lens' ServiceId (Maybe Text)

serviceId_type :: Lens' ServiceId (Maybe Text)

ServiceInfo

serviceInfo_state :: Lens' ServiceInfo (Maybe Text)

The service's state.

serviceInfo_startTime :: Lens' ServiceInfo (Maybe UTCTime)

The start time of the first segment that the service generated.

serviceInfo_root :: Lens' ServiceInfo (Maybe Bool)

Indicates that the service was the first service to process a request.

serviceInfo_responseTimeHistogram :: Lens' ServiceInfo (Maybe [HistogramEntry])

A histogram that maps the spread of service response times.

serviceInfo_durationHistogram :: Lens' ServiceInfo (Maybe [HistogramEntry])

A histogram that maps the spread of service durations.

serviceInfo_referenceId :: Lens' ServiceInfo (Maybe Int)

Identifier for the service. Unique within the service map.

serviceInfo_accountId :: Lens' ServiceInfo (Maybe Text)

Identifier of the Amazon Web Services account in which the service runs.

serviceInfo_names :: Lens' ServiceInfo (Maybe [Text])

A list of names for the service, including the canonical name.

serviceInfo_name :: Lens' ServiceInfo (Maybe Text)

The canonical name of the service.

serviceInfo_endTime :: Lens' ServiceInfo (Maybe UTCTime)

The end time of the last segment that the service generated.

serviceInfo_type :: Lens' ServiceInfo (Maybe Text)

The type of service.

Amazon Web Services Resource - The type of an Amazon Web Services resource. For example, AWS::EC2::Instance for an application running on Amazon EC2 or AWS::DynamoDB::Table for an Amazon DynamoDB table that the application used.

Amazon Web Services Service - The type of an Amazon Web Services service. For example, AWS::DynamoDB for downstream calls to Amazon DynamoDB that didn't target a specific table.

client - Represents the clients that sent requests to a root service.

remote - A downstream service of indeterminate type.

serviceInfo_edges :: Lens' ServiceInfo (Maybe [Edge])

Connections to downstream services.

serviceInfo_summaryStatistics :: Lens' ServiceInfo (Maybe ServiceStatistics)

Aggregated statistics for the service.

ServiceStatistics

serviceStatistics_faultStatistics :: Lens' ServiceStatistics (Maybe FaultStatistics)

Information about requests that failed with a 5xx Server Error status code.

serviceStatistics_okCount :: Lens' ServiceStatistics (Maybe Integer)

The number of requests that completed with a 2xx Success status code.

serviceStatistics_totalResponseTime :: Lens' ServiceStatistics (Maybe Double)

The aggregate response time of completed requests.

serviceStatistics_errorStatistics :: Lens' ServiceStatistics (Maybe ErrorStatistics)

Information about requests that failed with a 4xx Client Error status code.

serviceStatistics_totalCount :: Lens' ServiceStatistics (Maybe Integer)

The total number of completed requests.

Tag

tag_key :: Lens' Tag Text

A tag key, such as Stage or Name. A tag key cannot be empty. The key can be a maximum of 128 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /

tag_value :: Lens' Tag Text

An optional tag value, such as Production or test-only. The value can be a maximum of 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /

TelemetryRecord

telemetryRecord_segmentsReceivedCount :: Lens' TelemetryRecord (Maybe Int)

telemetryRecord_segmentsSentCount :: Lens' TelemetryRecord (Maybe Int)

telemetryRecord_segmentsSpilloverCount :: Lens' TelemetryRecord (Maybe Int)

telemetryRecord_segmentsRejectedCount :: Lens' TelemetryRecord (Maybe Int)

telemetryRecord_backendConnectionErrors :: Lens' TelemetryRecord (Maybe BackendConnectionErrors)

telemetryRecord_timestamp :: Lens' TelemetryRecord UTCTime

TimeSeriesServiceStatistics

timeSeriesServiceStatistics_serviceSummaryStatistics :: Lens' TimeSeriesServiceStatistics (Maybe ServiceStatistics)

Undocumented member.

timeSeriesServiceStatistics_responseTimeHistogram :: Lens' TimeSeriesServiceStatistics (Maybe [HistogramEntry])

The response time histogram for the selected entities.

timeSeriesServiceStatistics_edgeSummaryStatistics :: Lens' TimeSeriesServiceStatistics (Maybe EdgeStatistics)

Undocumented member.

timeSeriesServiceStatistics_serviceForecastStatistics :: Lens' TimeSeriesServiceStatistics (Maybe ForecastStatistics)

The forecasted high and low fault count values.

timeSeriesServiceStatistics_timestamp :: Lens' TimeSeriesServiceStatistics (Maybe UTCTime)

Timestamp of the window for which statistics are aggregated.

Trace

trace_limitExceeded :: Lens' Trace (Maybe Bool)

LimitExceeded is set to true when the trace has exceeded one of the defined quotas. For more information about quotas, see Amazon Web Services X-Ray endpoints and quotas.

trace_id :: Lens' Trace (Maybe Text)

The unique identifier for the request that generated the trace's segments and subsegments.

trace_segments :: Lens' Trace (Maybe [Segment])

Segment documents for the segments and subsegments that comprise the trace.

trace_duration :: Lens' Trace (Maybe Double)

The length of time in seconds between the start time of the root segment and the end time of the last segment that completed.

TraceSummary

traceSummary_annotations :: Lens' TraceSummary (Maybe (HashMap Text [ValueWithServiceIds]))

Annotations from the trace's segment documents.

traceSummary_hasThrottle :: Lens' TraceSummary (Maybe Bool)

One or more of the segment documents has a 429 throttling error.

traceSummary_users :: Lens' TraceSummary (Maybe [TraceUser])

Users from the trace's segment documents.

traceSummary_entryPoint :: Lens' TraceSummary (Maybe ServiceId)

The root of a trace.

traceSummary_hasFault :: Lens' TraceSummary (Maybe Bool)

The root segment document has a 500 series error.

traceSummary_serviceIds :: Lens' TraceSummary (Maybe [ServiceId])

Service IDs from the trace's segment documents.

traceSummary_matchedEventTime :: Lens' TraceSummary (Maybe UTCTime)

The matched time stamp of a defined event.

traceSummary_isPartial :: Lens' TraceSummary (Maybe Bool)

One or more of the segment documents is in progress.

traceSummary_errorRootCauses :: Lens' TraceSummary (Maybe [ErrorRootCause])

A collection of ErrorRootCause structures corresponding to the trace segments.

traceSummary_resourceARNs :: Lens' TraceSummary (Maybe [ResourceARNDetail])

A list of resource ARNs for any resource corresponding to the trace segments.

traceSummary_availabilityZones :: Lens' TraceSummary (Maybe [AvailabilityZoneDetail])

A list of Availability Zones for any zone corresponding to the trace segments.

traceSummary_instanceIds :: Lens' TraceSummary (Maybe [InstanceIdDetail])

A list of EC2 instance IDs for any instance corresponding to the trace segments.

traceSummary_responseTimeRootCauses :: Lens' TraceSummary (Maybe [ResponseTimeRootCause])

A collection of ResponseTimeRootCause structures corresponding to the trace segments.

traceSummary_hasError :: Lens' TraceSummary (Maybe Bool)

The root segment document has a 400 series error.

traceSummary_id :: Lens' TraceSummary (Maybe Text)

The unique identifier for the request that generated the trace's segments and subsegments.

traceSummary_http :: Lens' TraceSummary (Maybe Http)

Information about the HTTP request served by the trace.

traceSummary_revision :: Lens' TraceSummary (Maybe Int)

The revision number of a trace.

traceSummary_duration :: Lens' TraceSummary (Maybe Double)

The length of time in seconds between the start time of the root segment and the end time of the last segment that completed.

traceSummary_faultRootCauses :: Lens' TraceSummary (Maybe [FaultRootCause])

A collection of FaultRootCause structures corresponding to the trace segments.

traceSummary_responseTime :: Lens' TraceSummary (Maybe Double)

The length of time in seconds between the start and end times of the root segment. If the service performs work asynchronously, the response time measures the time before the response is sent to the user, while the duration measures the amount of time before the last traced activity completes.

TraceUser

traceUser_serviceIds :: Lens' TraceUser (Maybe [ServiceId])

Services that the user's request hit.

traceUser_userName :: Lens' TraceUser (Maybe Text)

The user's name.

UnprocessedStatistics

unprocessedStatistics_ruleName :: Lens' UnprocessedStatistics (Maybe Text)

The name of the sampling rule.

unprocessedStatistics_errorCode :: Lens' UnprocessedStatistics (Maybe Text)

The error code.

unprocessedStatistics_message :: Lens' UnprocessedStatistics (Maybe Text)

The error message.

UnprocessedTraceSegment

unprocessedTraceSegment_errorCode :: Lens' UnprocessedTraceSegment (Maybe Text)

The error that caused processing to fail.

unprocessedTraceSegment_id :: Lens' UnprocessedTraceSegment (Maybe Text)

The segment's ID.

unprocessedTraceSegment_message :: Lens' UnprocessedTraceSegment (Maybe Text)

The error message.

ValueWithServiceIds

valueWithServiceIds_serviceIds :: Lens' ValueWithServiceIds (Maybe [ServiceId])

Services to which the annotation applies.

valueWithServiceIds_annotationValue :: Lens' ValueWithServiceIds (Maybe AnnotationValue)

Values of the annotation.