Amazonka.XRay.Types.Edge

Description

Documentation

data Edge

Information about a connection between two services.

See: newEdge smart constructor.

Constructors

Edge'

Fields

startTime :: Maybe POSIX

The start time of the first segment on the edge.

aliases :: Maybe [Alias]

Aliases for the edge.

responseTimeHistogram :: Maybe [HistogramEntry]

A histogram that maps the spread of client response times on an edge.

referenceId :: Maybe Int

Identifier of the edge. Unique within a service map.

endTime :: Maybe POSIX

The end time of the last segment on the edge.

summaryStatistics :: Maybe EdgeStatistics

Response statistics for segments on the edge.

newEdge :: Edge

Create a value of Edge with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:Edge', edge_startTime - The start time of the first segment on the edge.

$sel:aliases:Edge', edge_aliases - Aliases for the edge.

$sel:responseTimeHistogram:Edge', edge_responseTimeHistogram - A histogram that maps the spread of client response times on an edge.

$sel:referenceId:Edge', edge_referenceId - Identifier of the edge. Unique within a service map.

$sel:endTime:Edge', edge_endTime - The end time of the last segment on the edge.

$sel:summaryStatistics:Edge', edge_summaryStatistics - Response statistics for segments on the edge.

edge_startTime :: Lens' Edge (Maybe UTCTime)

The start time of the first segment on the edge.

edge_aliases :: Lens' Edge (Maybe [Alias])

Aliases for the edge.

edge_responseTimeHistogram :: Lens' Edge (Maybe [HistogramEntry])

A histogram that maps the spread of client response times on an edge.

edge_referenceId :: Lens' Edge (Maybe Int)

Identifier of the edge. Unique within a service map.

edge_endTime :: Lens' Edge (Maybe UTCTime)

The end time of the last segment on the edge.

edge_summaryStatistics :: Lens' Edge (Maybe EdgeStatistics)

Response statistics for segments on the edge.