Amazonka.XRay.Types.ErrorRootCauseService

Description

Documentation

data ErrorRootCauseService

A collection of fields identifying the services in a trace summary error.

See: newErrorRootCauseService smart constructor.

Constructors

ErrorRootCauseService'

Fields

entityPath :: Maybe [ErrorRootCauseEntity]

The path of root cause entities found on the service.

accountId :: Maybe Text

The account ID associated to the service.

names :: Maybe [Text]

A collection of associated service names.

name :: Maybe Text

The service name.

inferred :: Maybe Bool

A Boolean value indicating if the service is inferred from the trace.

type' :: Maybe Text

The type associated to the service.

newErrorRootCauseService :: ErrorRootCauseService

Create a value of ErrorRootCauseService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entityPath:ErrorRootCauseService', errorRootCauseService_entityPath - The path of root cause entities found on the service.

$sel:accountId:ErrorRootCauseService', errorRootCauseService_accountId - The account ID associated to the service.

$sel:names:ErrorRootCauseService', errorRootCauseService_names - A collection of associated service names.

$sel:name:ErrorRootCauseService', errorRootCauseService_name - The service name.

$sel:inferred:ErrorRootCauseService', errorRootCauseService_inferred - A Boolean value indicating if the service is inferred from the trace.

$sel:type':ErrorRootCauseService', errorRootCauseService_type - The type associated to the service.

errorRootCauseService_entityPath :: Lens' ErrorRootCauseService (Maybe [ErrorRootCauseEntity])

The path of root cause entities found on the service.

errorRootCauseService_accountId :: Lens' ErrorRootCauseService (Maybe Text)

The account ID associated to the service.

errorRootCauseService_names :: Lens' ErrorRootCauseService (Maybe [Text])

A collection of associated service names.

errorRootCauseService_name :: Lens' ErrorRootCauseService (Maybe Text)

The service name.

errorRootCauseService_inferred :: Lens' ErrorRootCauseService (Maybe Bool)

A Boolean value indicating if the service is inferred from the trace.

errorRootCauseService_type :: Lens' ErrorRootCauseService (Maybe Text)

The type associated to the service.