Amazonka.XRay.Types.InsightImpactGraphEdge

Description

Documentation

data InsightImpactGraphEdge

The connection between two service in an insight impact graph.

See: newInsightImpactGraphEdge smart constructor.

Constructors

InsightImpactGraphEdge'

Fields

referenceId :: Maybe Int

Identifier of the edge. Unique within a service map.

newInsightImpactGraphEdge :: InsightImpactGraphEdge

Create a value of InsightImpactGraphEdge with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:referenceId:InsightImpactGraphEdge', insightImpactGraphEdge_referenceId - Identifier of the edge. Unique within a service map.

insightImpactGraphEdge_referenceId :: Lens' InsightImpactGraphEdge (Maybe Int)

Identifier of the edge. Unique within a service map.