Control.Monad.Trans.AWS

Description

A monad transformer built on top of functions from Network.AWS which encapsulates various common parameters, errors, and usage patterns.

Transformer

type AWS = AWST IO

A convenient alias for AWST IO.

data AWST m a

The transformer. This satisfies all of the constraints that the functions in this module require, such as providing MonadResource instances, and keeping track of the Env environment.

The MonadError instance for this transformer internally uses ExceptT to handle actions that result in an Error. For more information see sendCatch and paginateCatch.

type MonadAWS m = (MonadBaseControl IO m, MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

Provides an alias for shortening type signatures if preferred.

Note: requires the ConstraintKinds extension.

Running

runAWST :: MonadBaseControl IO m => Env -> AWST m a -> m (Either Error a)

Unwrap an AWST transformer, calling all of the registered ResourceT release actions.

Environment

data Env

The environment containing the parameters required to make AWS requests.

envAuth :: Lens' Env Auth

envRegion :: Lens' Env Region

envManager :: Lens' Env Manager

envLogger :: Lens' Env Logger

envRetry :: Lens' Env (Maybe RetryPolicy)

Creating the environment

newEnv :: (Functor m, MonadIO m) => Region -> Credentials -> Manager -> ExceptT String m Env

This creates a new environment without debug logging and uses getAuth to expand/discover the supplied Credentials.

Lenses such as envLogger can be used to modify the Env with a debug logger.

getEnv :: Region -> Credentials -> IO Env

Create a new environment without debug logging, creating a new Manager.

Any errors are thrown using error.

See: newEnv

Authentication credentials

Defaults

accessKey

Arguments

:: Text
AWS_ACCESS_KEY

Default access key environment variable.

secretKey

Arguments

:: Text
AWS_SECRET_KEY

Default secret key environment variable.

Specifying credentials

fromKeys :: AccessKey -> SecretKey -> Auth

Explicit access and secret keys.

fromSession :: AccessKey -> SecretKey -> SecurityToken -> Auth

A session containing the access key, secret key, and a security token.

Retrieving credentials

data Credentials

Determines how authentication information is retrieved.

Constructors

FromKeys AccessKey SecretKey
Explicit access and secret keys. Note: you can achieve the same result purely using fromKeys without having to use the impure getAuth.

FromSession AccessKey SecretKey SecurityToken
A session containing the access key, secret key, and a security token. Note: you can achieve the same result purely using fromSession without having to use the impure getAuth.

FromProfile Text
An IAM Profile name to lookup from the local EC2 instance-data.

FromEnv Text Text
Environment variables to lookup for the access and secret keys.

Discover
Attempt to read the default access and secret keys from the environment, falling back to the first available IAM profile if they are not set.

Note: This attempts to resolve http://instance-data rather than directly retrieving http://169.254.169.254 for IAM profile information to ensure the dns lookup terminates promptly if not running on EC2.

getAuth :: (Functor m, MonadIO m) => Manager -> Credentials -> ExceptT String m Auth

Retrieve authentication information using the specified Credentials style.

Logging

data LogLevel

Constructors

Info
Informational messages supplied by the user, not used by the library.

Debug
Info level + debug messages + non-streaming response bodies.

Trace
Debug level + potentially sensitive signing metadata.

type Logger = LogLevel -> Builder -> IO ()

newLogger :: MonadIO m => LogLevel -> Handle -> m Logger

This is a primitive logger which can be used to log messages to a Handle. A more sophisticated logging library such as tinylog or FastLogger should be used in production code.

logInfo :: (MonadIO m, MonadReader Env m, ToBuilder a) => a -> m ()

Use the supplied logger from envLogger to log info messages.

logDebug :: (MonadIO m, MonadReader Env m, ToBuilder a) => a -> m ()

Use the supplied logger from envLogger to log debug messages.

logTrace :: (MonadIO m, MonadReader Env m, ToBuilder a) => a -> m ()

Use the supplied logger from envLogger to log trace messages.

Regionalisation

data Region :: *

The sum of available AWS regions.

Constructors

Ireland
Europe / eu-west-1

Frankfurt
Europe / eu-central-1

Tokyo
Asia Pacific / ap-northeast-1

Singapore
Asia Pacific / ap-southeast-1

Sydney
Asia Pacific / ap-southeast-2

Beijing
China / cn-north-1

NorthVirginia
US / us-east-1

NorthCalifornia
US / us-west-1

Oregon
US / us-west-2

GovCloud
AWS GovCloud / us-gov-west-1

GovCloudFIPS
AWS GovCloud (FIPS 140-2) S3 Only / fips-us-gov-west-1

SaoPaulo
South America / sa-east-1

within :: MonadReader Env m => Region -> m a -> m a

Scope a monadic action within the specific Region.

Retries

once :: MonadReader Env m => m a -> m a

Scope a monadic action such that any potential retry logic for the Service is ignored.

Example: Any requests will at most be sent once.

Errors

type Error = ServiceError String

The top-level error type.

hoistEither :: (MonadError Error m, AWSError e) => Either e a -> m a

Hoist an Either throwing the Left case, and returning the Right.

throwAWSError :: (MonadError Error m, AWSError e) => e -> m a

verify :: (AWSError e, MonadError Error m) => Prism' e a -> e -> m ()

verifyWith :: (AWSError e, MonadError Error m) => Prism' e a -> (a -> Bool) -> e -> m ()

Requests

Synchronous

send :: (MonadCatch m, MonadResource m, MonadReader Env m, MonadError Error m, AWSRequest a) => a -> m (Rs a)

Send a data type which is an instance of AWSRequest, returning it's associated Rs response type.

This will throw any HTTPException or AWSServiceError returned by the service using the MonadError instance. In the case of AWST this will cause the internal ExceptT to short-circuit and return an Error in the Left case as the result of the computation.

See: sendCatch

send_ :: (MonadCatch m, MonadResource m, MonadReader Env m, MonadError Error m, AWSRequest a) => a -> m ()

A variant of send which discards any successful response.

See: send

sendCatch :: (MonadCatch m, MonadResource m, MonadReader Env m, AWSRequest a) => a -> m (Response a)

Send a data type which is an instance of AWSRequest, returning either the associated Rs response type in the success case, or the related service's Er type in the error case.

This includes HTTPExceptions, serialisation errors, and any service errors returned as part of the Response.

Note: Requests will be retried depending upon each service's respective strategy. This can be overriden using once or envRetry. Requests which contain streaming request bodies (such as S3's PutObject) are never considered for retries.

Paginated

paginate :: (MonadCatch m, MonadResource m, MonadReader Env m, MonadError Error m, AWSPager a) => a -> Source m (Rs a)

Send a data type which is an instance of AWSPager and paginate while there are more results as defined by the related service operation.

Errors will be handle identically to send.

Note: The ResumableSource will close when there are no more results or the ResourceT computation is unwrapped. See: runResourceT for more information.

See: paginateCatch

paginateCatch :: (MonadCatch m, MonadResource m, MonadReader Env m, AWSPager a) => a -> Source m (Response a)

Send a data type which is an instance of AWSPager and paginate over the associated Rs response type in the success case, or the related service's Er type in the error case.

Note: The ResumableSource will close when there are no more results or the ResourceT computation is unwrapped. See: runResourceT for more information.

Eventual consistency

await :: (MonadCatch m, MonadResource m, MonadReader Env m, MonadError Error m, AWSRequest a) => Wait a -> a -> m (Rs a)

Poll the API until a predfined condition is fulfilled using the supplied Wait specification from the respective service.

Any errors which are unhandled by the Wait specification during retries will be thrown in the same manner as send.

See: awaitCatch

Pre-signing URLs

presign

Arguments

:: (MonadIO m, MonadReader Env m, AWSRequest a, AWSPresigner (Sg (Sv a)))

=> a
Request to presign.

-> UTCTime
Signing time.

-> UTCTime
Expiry time.

-> m (Signed a (Sg (Sv a)))

Presign a URL with expiry to be used at a later time.

Note: Requires the service's signer to be an instance of AWSPresigner. Not all signing process support this.

Types

class ToBuilder a where

Minimal complete definition

Nothing

Methods

build :: a -> Builder

module Network.AWS.Types

module Network.AWS.Error