Network.AWS

Description

AWS Context

data AWS a

runAWS :: Credentials -> Bool -> AWS a -> IO (Either AWSError a)

Credentials

data Credentials

Constructors

FromKeys ByteString ByteString

FromRole ByteString

Regions

data Region

Constructors

NorthVirginia

NorthCalifornia

Oregon

Ireland

Singapore

Tokyo

Sydney

SaoPaulo

within :: Region -> AWS a -> AWS a

Run an AWS operation inside a specific Region.

getRegion :: AWS Region

Debugging

getDebug :: AWS Bool

whenDebug :: AWS () -> AWS ()

Synchronous Requests

send :: (Rq a, ToError (Er a)) => a -> AWS (Rs a)

Send a request and return the associated response type.

send_ :: (Rq a, ToError (Er a)) => a -> AWS ()

Send a request ignoring the succesful response but propagating errors.

sendCatch :: Rq a => a -> AWS (Either (Er a) (Rs a))

Asynchronous Requests

async :: AWS a -> AWS (Async (Either AWSError a))

sendAsync :: Rq a => a -> AWS (Async (Either AWSError (Either (Er a) (Rs a))))

wait :: Async (Either AWSError a) -> AWS a

wait_ :: Async (Either AWSError a) -> AWS ()

waitAsync :: ToError e => Async (Either AWSError (Either e a)) -> AWS a

waitAsync_ :: ToError e => Async (Either AWSError (Either e a)) -> AWS ()

Paginated Requests

paginate :: (Rq a, Pg a, ToError (Er a)) => a -> Producer' (Rs a) AWS ()

Create a pipes Producer which yields the initial and subsequent repsonses for requests that supported pagination.

paginateCatch :: (Rq a, Pg a, ToError (Er a)) => a -> Producer' (Either (Er a) (Rs a)) AWS ()

Errors

class ToError a where

Methods

toError :: a -> AWSError

data AWSError

Constructors

Err String

Ex SomeException

hoistError :: (MonadError e m, Error e) => Either e a -> m a

liftEitherT :: ToError e => EitherT e IO a -> AWS a

Types

data AvailabilityZone

Constructors

AZ

Fields

azRegion :: !Region

azSuffix :: !Char

data Body

Constructors

Strict ByteString

Streaming (InputStream ByteString)

Empty

data InstanceType

Constructors

T1_Micro

M1_Small

M1_Medium

M1_Large

M1_XLarge

M3_XLarge

M3_2XLarge

C3_XLarge

C3_2XLarge

C3_4XLarge

C3_8XLarge

C1_Medium

C1_XLarge

CC2_8XLarge

M2_XLarge

M2_2XLarge

M2_4XLarge

CR1_8XLarge

HI1_4XLarge

HS1_8XLarge

CG1_4XLarge

newtype Items a

Constructors

Items

Fields

items :: [a]

newtype Members a

Constructors

Members

Fields

members :: [a]