Network.AWS.AMP.CreateRuleGroupsNamespace

Description

Create a rule group namespace.

Creating a Request

data CreateRuleGroupsNamespace

Represents the input of a CreateRuleGroupsNamespace operation.

See: newCreateRuleGroupsNamespace smart constructor.

Constructors

CreateRuleGroupsNamespace'

Fields

clientToken :: Maybe Text

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

tags :: Maybe (HashMap Text Text)

Optional, user-provided tags for this rule groups namespace.

data' :: Base64

The namespace data that define the rule groups.

name :: Text

The rule groups namespace name.

workspaceId :: Text

The ID of the workspace in which to create the rule group namespace.

newCreateRuleGroupsNamespace

Arguments

:: ByteString
$sel:data':CreateRuleGroupsNamespace'

-> Text
$sel:name:CreateRuleGroupsNamespace'

-> Text
$sel:workspaceId:CreateRuleGroupsNamespace'

-> CreateRuleGroupsNamespace

Create a value of CreateRuleGroupsNamespace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateRuleGroupsNamespace', createRuleGroupsNamespace_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:tags:CreateRuleGroupsNamespace', createRuleGroupsNamespace_tags - Optional, user-provided tags for this rule groups namespace.

$sel:data':CreateRuleGroupsNamespace', createRuleGroupsNamespace_data - The namespace data that define the rule groups.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:name:CreateRuleGroupsNamespace', createRuleGroupsNamespace_name - The rule groups namespace name.

$sel:workspaceId:CreateRuleGroupsNamespace', createRuleGroupsNamespace_workspaceId - The ID of the workspace in which to create the rule group namespace.

Request Lenses

createRuleGroupsNamespace_clientToken :: Lens' CreateRuleGroupsNamespace (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

createRuleGroupsNamespace_tags :: Lens' CreateRuleGroupsNamespace (Maybe (HashMap Text Text))

Optional, user-provided tags for this rule groups namespace.

createRuleGroupsNamespace_data :: Lens' CreateRuleGroupsNamespace ByteString

The namespace data that define the rule groups.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

createRuleGroupsNamespace_name :: Lens' CreateRuleGroupsNamespace Text

The rule groups namespace name.

createRuleGroupsNamespace_workspaceId :: Lens' CreateRuleGroupsNamespace Text

The ID of the workspace in which to create the rule group namespace.

Destructuring the Response

data CreateRuleGroupsNamespaceResponse

Represents the output of a CreateRuleGroupsNamespace operation.

See: newCreateRuleGroupsNamespaceResponse smart constructor.

Constructors

CreateRuleGroupsNamespaceResponse'

Fields

tags :: Maybe (HashMap Text Text)

The tags of this rule groups namespace.

httpStatus :: Int

The response's http status code.

arn :: Text

The Amazon Resource Name (ARN) of this rule groups namespace.

name :: Text

The rule groups namespace name.

status :: RuleGroupsNamespaceStatus

The status of rule groups namespace.

newCreateRuleGroupsNamespaceResponse

Arguments

:: Int
$sel:httpStatus:CreateRuleGroupsNamespaceResponse'

-> Text
$sel:arn:CreateRuleGroupsNamespaceResponse'

-> Text
$sel:name:CreateRuleGroupsNamespace'

-> RuleGroupsNamespaceStatus
$sel:status:CreateRuleGroupsNamespaceResponse'

-> CreateRuleGroupsNamespaceResponse

Create a value of CreateRuleGroupsNamespaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateRuleGroupsNamespace', createRuleGroupsNamespaceResponse_tags - The tags of this rule groups namespace.

$sel:httpStatus:CreateRuleGroupsNamespaceResponse', createRuleGroupsNamespaceResponse_httpStatus - The response's http status code.

$sel:arn:CreateRuleGroupsNamespaceResponse', createRuleGroupsNamespaceResponse_arn - The Amazon Resource Name (ARN) of this rule groups namespace.

$sel:name:CreateRuleGroupsNamespace', createRuleGroupsNamespaceResponse_name - The rule groups namespace name.

$sel:status:CreateRuleGroupsNamespaceResponse', createRuleGroupsNamespaceResponse_status - The status of rule groups namespace.

Response Lenses

createRuleGroupsNamespaceResponse_tags :: Lens' CreateRuleGroupsNamespaceResponse (Maybe (HashMap Text Text))

The tags of this rule groups namespace.

createRuleGroupsNamespaceResponse_httpStatus :: Lens' CreateRuleGroupsNamespaceResponse Int

The response's http status code.

createRuleGroupsNamespaceResponse_arn :: Lens' CreateRuleGroupsNamespaceResponse Text

The Amazon Resource Name (ARN) of this rule groups namespace.

createRuleGroupsNamespaceResponse_name :: Lens' CreateRuleGroupsNamespaceResponse Text

The rule groups namespace name.

createRuleGroupsNamespaceResponse_status :: Lens' CreateRuleGroupsNamespaceResponse RuleGroupsNamespaceStatus

The status of rule groups namespace.