Network.AWS.AMP.CreateWorkspace

Description

Creates a new AMP workspace.

Creating a Request

data CreateWorkspace

Represents the input of a CreateWorkspace operation.

See: newCreateWorkspace smart constructor.

Constructors

CreateWorkspace'

Fields

clientToken :: Maybe Text

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

alias :: Maybe Text

An optional user-assigned alias for this workspace. This alias is for user reference and does not need to be unique.

tags :: Maybe (HashMap Text Text)

Optional, user-provided tags for this workspace.

newCreateWorkspace :: CreateWorkspace

Create a value of CreateWorkspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateWorkspace', createWorkspace_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:alias:CreateWorkspace', createWorkspace_alias - An optional user-assigned alias for this workspace. This alias is for user reference and does not need to be unique.

$sel:tags:CreateWorkspace', createWorkspace_tags - Optional, user-provided tags for this workspace.

Request Lenses

createWorkspace_clientToken :: Lens' CreateWorkspace (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

createWorkspace_alias :: Lens' CreateWorkspace (Maybe Text)

An optional user-assigned alias for this workspace. This alias is for user reference and does not need to be unique.

createWorkspace_tags :: Lens' CreateWorkspace (Maybe (HashMap Text Text))

Optional, user-provided tags for this workspace.

Destructuring the Response

data CreateWorkspaceResponse

Represents the output of a CreateWorkspace operation.

See: newCreateWorkspaceResponse smart constructor.

Constructors

CreateWorkspaceResponse'

Fields

tags :: Maybe (HashMap Text Text)

The tags of this workspace.

httpStatus :: Int

The response's http status code.

arn :: Text

The ARN of the workspace that was just created.

status :: WorkspaceStatus

The status of the workspace that was just created (usually CREATING).

workspaceId :: Text

The generated ID of the workspace that was just created.

newCreateWorkspaceResponse

Arguments

:: Int
$sel:httpStatus:CreateWorkspaceResponse'

-> Text
$sel:arn:CreateWorkspaceResponse'

-> WorkspaceStatus
$sel:status:CreateWorkspaceResponse'

-> Text
$sel:workspaceId:CreateWorkspaceResponse'

-> CreateWorkspaceResponse

Create a value of CreateWorkspaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateWorkspace', createWorkspaceResponse_tags - The tags of this workspace.

$sel:httpStatus:CreateWorkspaceResponse', createWorkspaceResponse_httpStatus - The response's http status code.

$sel:arn:CreateWorkspaceResponse', createWorkspaceResponse_arn - The ARN of the workspace that was just created.

$sel:status:CreateWorkspaceResponse', createWorkspaceResponse_status - The status of the workspace that was just created (usually CREATING).

$sel:workspaceId:CreateWorkspaceResponse', createWorkspaceResponse_workspaceId - The generated ID of the workspace that was just created.

Response Lenses

createWorkspaceResponse_tags :: Lens' CreateWorkspaceResponse (Maybe (HashMap Text Text))

The tags of this workspace.

createWorkspaceResponse_httpStatus :: Lens' CreateWorkspaceResponse Int

The response's http status code.

createWorkspaceResponse_arn :: Lens' CreateWorkspaceResponse Text

The ARN of the workspace that was just created.

createWorkspaceResponse_status :: Lens' CreateWorkspaceResponse WorkspaceStatus

The status of the workspace that was just created (usually CREATING).

createWorkspaceResponse_workspaceId :: Lens' CreateWorkspaceResponse Text

The generated ID of the workspace that was just created.