Network.AWS.AMP.DescribeWorkspace

Description

Describes an existing AMP workspace.

Creating a Request

data DescribeWorkspace

Represents the input of a DescribeWorkspace operation.

See: newDescribeWorkspace smart constructor.

Constructors

DescribeWorkspace'

Fields

workspaceId :: Text

The ID of the workspace to describe.

newDescribeWorkspace

Arguments

:: Text
$sel:workspaceId:DescribeWorkspace'

-> DescribeWorkspace

Create a value of DescribeWorkspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceId:DescribeWorkspace', describeWorkspace_workspaceId - The ID of the workspace to describe.

Request Lenses

describeWorkspace_workspaceId :: Lens' DescribeWorkspace Text

The ID of the workspace to describe.

Destructuring the Response

data DescribeWorkspaceResponse

Represents the output of a DescribeWorkspace operation.

See: newDescribeWorkspaceResponse smart constructor.

Constructors

DescribeWorkspaceResponse'

Fields

httpStatus :: Int

The response's http status code.

workspace :: WorkspaceDescription

The properties of the selected workspace.

newDescribeWorkspaceResponse

Arguments

:: Int
$sel:httpStatus:DescribeWorkspaceResponse'

-> WorkspaceDescription
$sel:workspace:DescribeWorkspaceResponse'

-> DescribeWorkspaceResponse

Create a value of DescribeWorkspaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeWorkspaceResponse', describeWorkspaceResponse_httpStatus - The response's http status code.

$sel:workspace:DescribeWorkspaceResponse', describeWorkspaceResponse_workspace - The properties of the selected workspace.

Response Lenses

describeWorkspaceResponse_httpStatus :: Lens' DescribeWorkspaceResponse Int

The response's http status code.

describeWorkspaceResponse_workspace :: Lens' DescribeWorkspaceResponse WorkspaceDescription

The properties of the selected workspace.