Network.AWS.AMP.ListWorkspaces

Description

Lists all AMP workspaces, including workspaces being created or deleted.

This operation returns paginated results.

Creating a Request

data ListWorkspaces

Represents the input of a ListWorkspaces operation.

See: newListWorkspaces smart constructor.

Constructors

ListWorkspaces'

Fields

alias :: Maybe Text

Optional filter for workspace alias. Only the workspaces with aliases that begin with this value will be returned.

nextToken :: Maybe Text

Pagination token to request the next page in a paginated list. This token is obtained from the output of the previous ListWorkspaces request.

maxResults :: Maybe Natural

Maximum results to return in response (default=100, maximum=1000).

newListWorkspaces :: ListWorkspaces

Create a value of ListWorkspaces with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alias:ListWorkspaces', listWorkspaces_alias - Optional filter for workspace alias. Only the workspaces with aliases that begin with this value will be returned.

$sel:nextToken:ListWorkspaces', listWorkspaces_nextToken - Pagination token to request the next page in a paginated list. This token is obtained from the output of the previous ListWorkspaces request.

$sel:maxResults:ListWorkspaces', listWorkspaces_maxResults - Maximum results to return in response (default=100, maximum=1000).

Request Lenses

listWorkspaces_alias :: Lens' ListWorkspaces (Maybe Text)

Optional filter for workspace alias. Only the workspaces with aliases that begin with this value will be returned.

listWorkspaces_nextToken :: Lens' ListWorkspaces (Maybe Text)

Pagination token to request the next page in a paginated list. This token is obtained from the output of the previous ListWorkspaces request.

listWorkspaces_maxResults :: Lens' ListWorkspaces (Maybe Natural)

Maximum results to return in response (default=100, maximum=1000).

Destructuring the Response

data ListWorkspacesResponse

Represents the output of a ListWorkspaces operation.

See: newListWorkspacesResponse smart constructor.

Constructors

ListWorkspacesResponse'

Fields

nextToken :: Maybe Text

Pagination token to use when requesting the next page in this list.

httpStatus :: Int

The response's http status code.

workspaces :: [WorkspaceSummary]

The list of existing workspaces, including those undergoing creation or deletion.

newListWorkspacesResponse

Arguments

:: Int
$sel:httpStatus:ListWorkspacesResponse'

-> ListWorkspacesResponse

Create a value of ListWorkspacesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorkspaces', listWorkspacesResponse_nextToken - Pagination token to use when requesting the next page in this list.

$sel:httpStatus:ListWorkspacesResponse', listWorkspacesResponse_httpStatus - The response's http status code.

$sel:workspaces:ListWorkspacesResponse', listWorkspacesResponse_workspaces - The list of existing workspaces, including those undergoing creation or deletion.

Response Lenses

listWorkspacesResponse_nextToken :: Lens' ListWorkspacesResponse (Maybe Text)

Pagination token to use when requesting the next page in this list.

listWorkspacesResponse_httpStatus :: Lens' ListWorkspacesResponse Int

The response's http status code.

listWorkspacesResponse_workspaces :: Lens' ListWorkspacesResponse [WorkspaceSummary]

The list of existing workspaces, including those undergoing creation or deletion.