Network.AWS.AMP.Types

Description

Service Configuration

defaultService :: Service

API version 2020-08-01 of the Amazon Prometheus Service SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

User does not have sufficient access to perform this action.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Updating or deleting a resource can cause an inconsistent state.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Request would cause a service quota to be exceeded.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Request was denied due to request throttling.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Unexpected error during processing of request.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Request references a resource which does not exist.

AlertManagerDefinitionStatusCode

newtype AlertManagerDefinitionStatusCode

State of an alert manager definition.

Constructors

AlertManagerDefinitionStatusCode'

Fields

fromAlertManagerDefinitionStatusCode :: Text

Bundled Patterns

pattern AlertManagerDefinitionStatusCode_ACTIVE :: AlertManagerDefinitionStatusCode

pattern AlertManagerDefinitionStatusCode_CREATING :: AlertManagerDefinitionStatusCode

pattern AlertManagerDefinitionStatusCode_CREATION_FAILED :: AlertManagerDefinitionStatusCode

pattern AlertManagerDefinitionStatusCode_DELETING :: AlertManagerDefinitionStatusCode

pattern AlertManagerDefinitionStatusCode_UPDATE_FAILED :: AlertManagerDefinitionStatusCode

pattern AlertManagerDefinitionStatusCode_UPDATING :: AlertManagerDefinitionStatusCode

RuleGroupsNamespaceStatusCode

newtype RuleGroupsNamespaceStatusCode

State of a namespace.

Constructors

RuleGroupsNamespaceStatusCode'

Fields

fromRuleGroupsNamespaceStatusCode :: Text

Bundled Patterns

pattern RuleGroupsNamespaceStatusCode_ACTIVE :: RuleGroupsNamespaceStatusCode

pattern RuleGroupsNamespaceStatusCode_CREATING :: RuleGroupsNamespaceStatusCode

pattern RuleGroupsNamespaceStatusCode_CREATION_FAILED :: RuleGroupsNamespaceStatusCode

pattern RuleGroupsNamespaceStatusCode_DELETING :: RuleGroupsNamespaceStatusCode

pattern RuleGroupsNamespaceStatusCode_UPDATE_FAILED :: RuleGroupsNamespaceStatusCode

pattern RuleGroupsNamespaceStatusCode_UPDATING :: RuleGroupsNamespaceStatusCode

WorkspaceStatusCode

newtype WorkspaceStatusCode

State of a workspace.

Constructors

WorkspaceStatusCode'

Fields

fromWorkspaceStatusCode :: Text

Bundled Patterns

pattern WorkspaceStatusCode_ACTIVE :: WorkspaceStatusCode

pattern WorkspaceStatusCode_CREATING :: WorkspaceStatusCode

pattern WorkspaceStatusCode_CREATION_FAILED :: WorkspaceStatusCode

pattern WorkspaceStatusCode_DELETING :: WorkspaceStatusCode

pattern WorkspaceStatusCode_UPDATING :: WorkspaceStatusCode

AlertManagerDefinitionDescription

data AlertManagerDefinitionDescription

Represents the properties of an alert manager definition.

See: newAlertManagerDefinitionDescription smart constructor.

Constructors

AlertManagerDefinitionDescription'

Fields

createdAt :: POSIX

The time when the alert manager definition was created.

data' :: Base64

The alert manager definition.

modifiedAt :: POSIX

The time when the alert manager definition was modified.

status :: AlertManagerDefinitionStatus

The status of alert manager definition.

newAlertManagerDefinitionDescription

Arguments

:: UTCTime
$sel:createdAt:AlertManagerDefinitionDescription'

-> ByteString
$sel:data':AlertManagerDefinitionDescription'

-> UTCTime
$sel:modifiedAt:AlertManagerDefinitionDescription'

-> AlertManagerDefinitionStatus
$sel:status:AlertManagerDefinitionDescription'

-> AlertManagerDefinitionDescription

Create a value of AlertManagerDefinitionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:AlertManagerDefinitionDescription', alertManagerDefinitionDescription_createdAt - The time when the alert manager definition was created.

$sel:data':AlertManagerDefinitionDescription', alertManagerDefinitionDescription_data - The alert manager definition.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:modifiedAt:AlertManagerDefinitionDescription', alertManagerDefinitionDescription_modifiedAt - The time when the alert manager definition was modified.

$sel:status:AlertManagerDefinitionDescription', alertManagerDefinitionDescription_status - The status of alert manager definition.

alertManagerDefinitionDescription_createdAt :: Lens' AlertManagerDefinitionDescription UTCTime

The time when the alert manager definition was created.

alertManagerDefinitionDescription_data :: Lens' AlertManagerDefinitionDescription ByteString

The alert manager definition.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

alertManagerDefinitionDescription_modifiedAt :: Lens' AlertManagerDefinitionDescription UTCTime

The time when the alert manager definition was modified.

alertManagerDefinitionDescription_status :: Lens' AlertManagerDefinitionDescription AlertManagerDefinitionStatus

The status of alert manager definition.

AlertManagerDefinitionStatus

data AlertManagerDefinitionStatus

Represents the status of a definition.

See: newAlertManagerDefinitionStatus smart constructor.

Constructors

AlertManagerDefinitionStatus'

Fields

statusReason :: Maybe Text

The reason for failure if any.

statusCode :: AlertManagerDefinitionStatusCode

Status code of this definition.

newAlertManagerDefinitionStatus

Arguments

:: AlertManagerDefinitionStatusCode
$sel:statusCode:AlertManagerDefinitionStatus'

-> AlertManagerDefinitionStatus

Create a value of AlertManagerDefinitionStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusReason:AlertManagerDefinitionStatus', alertManagerDefinitionStatus_statusReason - The reason for failure if any.

$sel:statusCode:AlertManagerDefinitionStatus', alertManagerDefinitionStatus_statusCode - Status code of this definition.

alertManagerDefinitionStatus_statusReason :: Lens' AlertManagerDefinitionStatus (Maybe Text)

The reason for failure if any.

alertManagerDefinitionStatus_statusCode :: Lens' AlertManagerDefinitionStatus AlertManagerDefinitionStatusCode

Status code of this definition.

RuleGroupsNamespaceDescription

data RuleGroupsNamespaceDescription

Represents a description of the rule groups namespace.

See: newRuleGroupsNamespaceDescription smart constructor.

Constructors

RuleGroupsNamespaceDescription'

Fields

tags :: Maybe (HashMap Text Text)

The tags of this rule groups namespace.

arn :: Text

The Amazon Resource Name (ARN) of this rule groups namespace.

createdAt :: POSIX

The time when the rule groups namespace was created.

data' :: Base64

The rule groups namespace data.

modifiedAt :: POSIX

The time when the rule groups namespace was modified.

name :: Text

The rule groups namespace name.

status :: RuleGroupsNamespaceStatus

The status of rule groups namespace.

newRuleGroupsNamespaceDescription

Arguments

:: Text
$sel:arn:RuleGroupsNamespaceDescription'

-> UTCTime
$sel:createdAt:RuleGroupsNamespaceDescription'

-> ByteString
$sel:data':RuleGroupsNamespaceDescription'

-> UTCTime
$sel:modifiedAt:RuleGroupsNamespaceDescription'

-> Text
$sel:name:RuleGroupsNamespaceDescription'

-> RuleGroupsNamespaceStatus
$sel:status:RuleGroupsNamespaceDescription'

-> RuleGroupsNamespaceDescription

Create a value of RuleGroupsNamespaceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_tags - The tags of this rule groups namespace.

$sel:arn:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_arn - The Amazon Resource Name (ARN) of this rule groups namespace.

$sel:createdAt:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_createdAt - The time when the rule groups namespace was created.

$sel:data':RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_data - The rule groups namespace data.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:modifiedAt:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_modifiedAt - The time when the rule groups namespace was modified.

$sel:name:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_name - The rule groups namespace name.

$sel:status:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_status - The status of rule groups namespace.

ruleGroupsNamespaceDescription_tags :: Lens' RuleGroupsNamespaceDescription (Maybe (HashMap Text Text))

The tags of this rule groups namespace.

ruleGroupsNamespaceDescription_arn :: Lens' RuleGroupsNamespaceDescription Text

The Amazon Resource Name (ARN) of this rule groups namespace.

ruleGroupsNamespaceDescription_createdAt :: Lens' RuleGroupsNamespaceDescription UTCTime

The time when the rule groups namespace was created.

ruleGroupsNamespaceDescription_data :: Lens' RuleGroupsNamespaceDescription ByteString

The rule groups namespace data.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

ruleGroupsNamespaceDescription_modifiedAt :: Lens' RuleGroupsNamespaceDescription UTCTime

The time when the rule groups namespace was modified.

ruleGroupsNamespaceDescription_name :: Lens' RuleGroupsNamespaceDescription Text

The rule groups namespace name.

ruleGroupsNamespaceDescription_status :: Lens' RuleGroupsNamespaceDescription RuleGroupsNamespaceStatus

The status of rule groups namespace.

RuleGroupsNamespaceStatus

data RuleGroupsNamespaceStatus

Represents the status of a namespace.

See: newRuleGroupsNamespaceStatus smart constructor.

Constructors

RuleGroupsNamespaceStatus'

Fields

statusReason :: Maybe Text

The reason for failure if any.

statusCode :: RuleGroupsNamespaceStatusCode

Status code of this namespace.

newRuleGroupsNamespaceStatus

Arguments

:: RuleGroupsNamespaceStatusCode
$sel:statusCode:RuleGroupsNamespaceStatus'

-> RuleGroupsNamespaceStatus

Create a value of RuleGroupsNamespaceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusReason:RuleGroupsNamespaceStatus', ruleGroupsNamespaceStatus_statusReason - The reason for failure if any.

$sel:statusCode:RuleGroupsNamespaceStatus', ruleGroupsNamespaceStatus_statusCode - Status code of this namespace.

ruleGroupsNamespaceStatus_statusReason :: Lens' RuleGroupsNamespaceStatus (Maybe Text)

The reason for failure if any.

ruleGroupsNamespaceStatus_statusCode :: Lens' RuleGroupsNamespaceStatus RuleGroupsNamespaceStatusCode

Status code of this namespace.

RuleGroupsNamespaceSummary

data RuleGroupsNamespaceSummary

Represents a summary of the rule groups namespace.

See: newRuleGroupsNamespaceSummary smart constructor.

Constructors

RuleGroupsNamespaceSummary'

Fields

tags :: Maybe (HashMap Text Text)

The tags of this rule groups namespace.

arn :: Text

The Amazon Resource Name (ARN) of this rule groups namespace.

createdAt :: POSIX

The time when the rule groups namespace was created.

modifiedAt :: POSIX

The time when the rule groups namespace was modified.

name :: Text

The rule groups namespace name.

status :: RuleGroupsNamespaceStatus

The status of rule groups namespace.

newRuleGroupsNamespaceSummary

Arguments

:: Text
$sel:arn:RuleGroupsNamespaceSummary'

-> UTCTime
$sel:createdAt:RuleGroupsNamespaceSummary'

-> UTCTime
$sel:modifiedAt:RuleGroupsNamespaceSummary'

-> Text
$sel:name:RuleGroupsNamespaceSummary'

-> RuleGroupsNamespaceStatus
$sel:status:RuleGroupsNamespaceSummary'

-> RuleGroupsNamespaceSummary

Create a value of RuleGroupsNamespaceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:RuleGroupsNamespaceSummary', ruleGroupsNamespaceSummary_tags - The tags of this rule groups namespace.

$sel:arn:RuleGroupsNamespaceSummary', ruleGroupsNamespaceSummary_arn - The Amazon Resource Name (ARN) of this rule groups namespace.

$sel:createdAt:RuleGroupsNamespaceSummary', ruleGroupsNamespaceSummary_createdAt - The time when the rule groups namespace was created.

$sel:modifiedAt:RuleGroupsNamespaceSummary', ruleGroupsNamespaceSummary_modifiedAt - The time when the rule groups namespace was modified.

$sel:name:RuleGroupsNamespaceSummary', ruleGroupsNamespaceSummary_name - The rule groups namespace name.

$sel:status:RuleGroupsNamespaceSummary', ruleGroupsNamespaceSummary_status - The status of rule groups namespace.

ruleGroupsNamespaceSummary_tags :: Lens' RuleGroupsNamespaceSummary (Maybe (HashMap Text Text))

The tags of this rule groups namespace.

ruleGroupsNamespaceSummary_arn :: Lens' RuleGroupsNamespaceSummary Text

The Amazon Resource Name (ARN) of this rule groups namespace.

ruleGroupsNamespaceSummary_createdAt :: Lens' RuleGroupsNamespaceSummary UTCTime

The time when the rule groups namespace was created.

ruleGroupsNamespaceSummary_modifiedAt :: Lens' RuleGroupsNamespaceSummary UTCTime

The time when the rule groups namespace was modified.

ruleGroupsNamespaceSummary_name :: Lens' RuleGroupsNamespaceSummary Text

The rule groups namespace name.

ruleGroupsNamespaceSummary_status :: Lens' RuleGroupsNamespaceSummary RuleGroupsNamespaceStatus

The status of rule groups namespace.

WorkspaceDescription

data WorkspaceDescription

Represents the properties of a workspace.

See: newWorkspaceDescription smart constructor.

Constructors

WorkspaceDescription'

Fields

alias :: Maybe Text

Alias of this workspace.

prometheusEndpoint :: Maybe Text

Prometheus endpoint URI.

tags :: Maybe (HashMap Text Text)

The tags of this workspace.

arn :: Text

The Amazon Resource Name (ARN) of this workspace.

createdAt :: POSIX

The time when the workspace was created.

status :: WorkspaceStatus

The status of this workspace.

workspaceId :: Text

Unique string identifying this workspace.

newWorkspaceDescription

Arguments

:: Text
$sel:arn:WorkspaceDescription'

-> UTCTime
$sel:createdAt:WorkspaceDescription'

-> WorkspaceStatus
$sel:status:WorkspaceDescription'

-> Text
$sel:workspaceId:WorkspaceDescription'

-> WorkspaceDescription

Create a value of WorkspaceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alias:WorkspaceDescription', workspaceDescription_alias - Alias of this workspace.

$sel:prometheusEndpoint:WorkspaceDescription', workspaceDescription_prometheusEndpoint - Prometheus endpoint URI.

$sel:tags:WorkspaceDescription', workspaceDescription_tags - The tags of this workspace.

$sel:arn:WorkspaceDescription', workspaceDescription_arn - The Amazon Resource Name (ARN) of this workspace.

$sel:createdAt:WorkspaceDescription', workspaceDescription_createdAt - The time when the workspace was created.

$sel:status:WorkspaceDescription', workspaceDescription_status - The status of this workspace.

$sel:workspaceId:WorkspaceDescription', workspaceDescription_workspaceId - Unique string identifying this workspace.

workspaceDescription_alias :: Lens' WorkspaceDescription (Maybe Text)

Alias of this workspace.

workspaceDescription_prometheusEndpoint :: Lens' WorkspaceDescription (Maybe Text)

Prometheus endpoint URI.

workspaceDescription_tags :: Lens' WorkspaceDescription (Maybe (HashMap Text Text))

The tags of this workspace.

workspaceDescription_arn :: Lens' WorkspaceDescription Text

The Amazon Resource Name (ARN) of this workspace.

workspaceDescription_createdAt :: Lens' WorkspaceDescription UTCTime

The time when the workspace was created.

workspaceDescription_status :: Lens' WorkspaceDescription WorkspaceStatus

The status of this workspace.

workspaceDescription_workspaceId :: Lens' WorkspaceDescription Text

Unique string identifying this workspace.

WorkspaceStatus

data WorkspaceStatus

Represents the status of a workspace.

See: newWorkspaceStatus smart constructor.

Constructors

WorkspaceStatus'

Fields

statusCode :: WorkspaceStatusCode

Status code of this workspace.

newWorkspaceStatus

Arguments

:: WorkspaceStatusCode
$sel:statusCode:WorkspaceStatus'

-> WorkspaceStatus

Create a value of WorkspaceStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusCode:WorkspaceStatus', workspaceStatus_statusCode - Status code of this workspace.

workspaceStatus_statusCode :: Lens' WorkspaceStatus WorkspaceStatusCode

Status code of this workspace.

WorkspaceSummary

data WorkspaceSummary

Represents a summary of the properties of a workspace.

See: newWorkspaceSummary smart constructor.

Constructors

WorkspaceSummary'

Fields

alias :: Maybe Text

Alias of this workspace.

tags :: Maybe (HashMap Text Text)

The tags of this workspace.

arn :: Text

The AmazonResourceName of this workspace.

createdAt :: POSIX

The time when the workspace was created.

status :: WorkspaceStatus

The status of this workspace.

workspaceId :: Text

Unique string identifying this workspace.

newWorkspaceSummary

Arguments

:: Text
$sel:arn:WorkspaceSummary'

-> UTCTime
$sel:createdAt:WorkspaceSummary'

-> WorkspaceStatus
$sel:status:WorkspaceSummary'

-> Text
$sel:workspaceId:WorkspaceSummary'

-> WorkspaceSummary

Create a value of WorkspaceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alias:WorkspaceSummary', workspaceSummary_alias - Alias of this workspace.

$sel:tags:WorkspaceSummary', workspaceSummary_tags - The tags of this workspace.

$sel:arn:WorkspaceSummary', workspaceSummary_arn - The AmazonResourceName of this workspace.

$sel:createdAt:WorkspaceSummary', workspaceSummary_createdAt - The time when the workspace was created.

$sel:status:WorkspaceSummary', workspaceSummary_status - The status of this workspace.

$sel:workspaceId:WorkspaceSummary', workspaceSummary_workspaceId - Unique string identifying this workspace.

workspaceSummary_alias :: Lens' WorkspaceSummary (Maybe Text)

Alias of this workspace.

workspaceSummary_tags :: Lens' WorkspaceSummary (Maybe (HashMap Text Text))

The tags of this workspace.

workspaceSummary_arn :: Lens' WorkspaceSummary Text

The AmazonResourceName of this workspace.

workspaceSummary_createdAt :: Lens' WorkspaceSummary UTCTime

The time when the workspace was created.

workspaceSummary_status :: Lens' WorkspaceSummary WorkspaceStatus

The status of this workspace.

workspaceSummary_workspaceId :: Lens' WorkspaceSummary Text

Unique string identifying this workspace.