Network.AWS.AMP.Types.AlertManagerDefinitionDescription

Description

Documentation

data AlertManagerDefinitionDescription

Represents the properties of an alert manager definition.

See: newAlertManagerDefinitionDescription smart constructor.

Constructors

AlertManagerDefinitionDescription'

Fields

createdAt :: POSIX

The time when the alert manager definition was created.

data' :: Base64

The alert manager definition.

modifiedAt :: POSIX

The time when the alert manager definition was modified.

status :: AlertManagerDefinitionStatus

The status of alert manager definition.

newAlertManagerDefinitionDescription

Arguments

:: UTCTime
$sel:createdAt:AlertManagerDefinitionDescription'

-> ByteString
$sel:data':AlertManagerDefinitionDescription'

-> UTCTime
$sel:modifiedAt:AlertManagerDefinitionDescription'

-> AlertManagerDefinitionStatus
$sel:status:AlertManagerDefinitionDescription'

-> AlertManagerDefinitionDescription

Create a value of AlertManagerDefinitionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:AlertManagerDefinitionDescription', alertManagerDefinitionDescription_createdAt - The time when the alert manager definition was created.

$sel:data':AlertManagerDefinitionDescription', alertManagerDefinitionDescription_data - The alert manager definition.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:modifiedAt:AlertManagerDefinitionDescription', alertManagerDefinitionDescription_modifiedAt - The time when the alert manager definition was modified.

$sel:status:AlertManagerDefinitionDescription', alertManagerDefinitionDescription_status - The status of alert manager definition.

alertManagerDefinitionDescription_createdAt :: Lens' AlertManagerDefinitionDescription UTCTime

The time when the alert manager definition was created.

alertManagerDefinitionDescription_data :: Lens' AlertManagerDefinitionDescription ByteString

The alert manager definition.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

alertManagerDefinitionDescription_modifiedAt :: Lens' AlertManagerDefinitionDescription UTCTime

The time when the alert manager definition was modified.

alertManagerDefinitionDescription_status :: Lens' AlertManagerDefinitionDescription AlertManagerDefinitionStatus

The status of alert manager definition.