Network.AWS.AMP.Types.RuleGroupsNamespaceDescription

Description

Documentation

data RuleGroupsNamespaceDescription

Represents a description of the rule groups namespace.

See: newRuleGroupsNamespaceDescription smart constructor.

Constructors

RuleGroupsNamespaceDescription'

Fields

tags :: Maybe (HashMap Text Text)

The tags of this rule groups namespace.

arn :: Text

The Amazon Resource Name (ARN) of this rule groups namespace.

createdAt :: POSIX

The time when the rule groups namespace was created.

data' :: Base64

The rule groups namespace data.

modifiedAt :: POSIX

The time when the rule groups namespace was modified.

name :: Text

The rule groups namespace name.

status :: RuleGroupsNamespaceStatus

The status of rule groups namespace.

newRuleGroupsNamespaceDescription

Arguments

:: Text
$sel:arn:RuleGroupsNamespaceDescription'

-> UTCTime
$sel:createdAt:RuleGroupsNamespaceDescription'

-> ByteString
$sel:data':RuleGroupsNamespaceDescription'

-> UTCTime
$sel:modifiedAt:RuleGroupsNamespaceDescription'

-> Text
$sel:name:RuleGroupsNamespaceDescription'

-> RuleGroupsNamespaceStatus
$sel:status:RuleGroupsNamespaceDescription'

-> RuleGroupsNamespaceDescription

Create a value of RuleGroupsNamespaceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_tags - The tags of this rule groups namespace.

$sel:arn:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_arn - The Amazon Resource Name (ARN) of this rule groups namespace.

$sel:createdAt:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_createdAt - The time when the rule groups namespace was created.

$sel:data':RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_data - The rule groups namespace data.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:modifiedAt:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_modifiedAt - The time when the rule groups namespace was modified.

$sel:name:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_name - The rule groups namespace name.

$sel:status:RuleGroupsNamespaceDescription', ruleGroupsNamespaceDescription_status - The status of rule groups namespace.

ruleGroupsNamespaceDescription_tags :: Lens' RuleGroupsNamespaceDescription (Maybe (HashMap Text Text))

The tags of this rule groups namespace.

ruleGroupsNamespaceDescription_arn :: Lens' RuleGroupsNamespaceDescription Text

The Amazon Resource Name (ARN) of this rule groups namespace.

ruleGroupsNamespaceDescription_createdAt :: Lens' RuleGroupsNamespaceDescription UTCTime

The time when the rule groups namespace was created.

ruleGroupsNamespaceDescription_data :: Lens' RuleGroupsNamespaceDescription ByteString

The rule groups namespace data.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

ruleGroupsNamespaceDescription_modifiedAt :: Lens' RuleGroupsNamespaceDescription UTCTime

The time when the rule groups namespace was modified.

ruleGroupsNamespaceDescription_name :: Lens' RuleGroupsNamespaceDescription Text

The rule groups namespace name.

ruleGroupsNamespaceDescription_status :: Lens' RuleGroupsNamespaceDescription RuleGroupsNamespaceStatus

The status of rule groups namespace.