Network.AWS.AMP.UpdateWorkspaceAlias

Description

Updates an AMP workspace alias.

Creating a Request

data UpdateWorkspaceAlias

Represents the input of an UpdateWorkspaceAlias operation.

See: newUpdateWorkspaceAlias smart constructor.

Constructors

UpdateWorkspaceAlias'

Fields

clientToken :: Maybe Text

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

alias :: Maybe Text

The new alias of the workspace.

workspaceId :: Text

The ID of the workspace being updated.

newUpdateWorkspaceAlias

Arguments

:: Text
$sel:workspaceId:UpdateWorkspaceAlias'

-> UpdateWorkspaceAlias

Create a value of UpdateWorkspaceAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateWorkspaceAlias', updateWorkspaceAlias_clientToken - Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:alias:UpdateWorkspaceAlias', updateWorkspaceAlias_alias - The new alias of the workspace.

$sel:workspaceId:UpdateWorkspaceAlias', updateWorkspaceAlias_workspaceId - The ID of the workspace being updated.

Request Lenses

updateWorkspaceAlias_clientToken :: Lens' UpdateWorkspaceAlias (Maybe Text)

Optional, unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

updateWorkspaceAlias_alias :: Lens' UpdateWorkspaceAlias (Maybe Text)

The new alias of the workspace.

updateWorkspaceAlias_workspaceId :: Lens' UpdateWorkspaceAlias Text

The ID of the workspace being updated.

Destructuring the Response

data UpdateWorkspaceAliasResponse

See: newUpdateWorkspaceAliasResponse smart constructor.

Constructors

UpdateWorkspaceAliasResponse'

newUpdateWorkspaceAliasResponse :: UpdateWorkspaceAliasResponse

Create a value of UpdateWorkspaceAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.