Network.AWS.APIGateway.CreateApiKey

Description

Create an ApiKey resource.

AWS CLI

Creating a Request

data CreateApiKey

Request to create an ApiKey resource.

See: newCreateApiKey smart constructor.

Constructors

CreateApiKey'

Fields

enabled :: Maybe Bool

Specifies whether the ApiKey can be used by callers.

value :: Maybe Text

Specifies a value of the API key.

customerId :: Maybe Text

An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

generateDistinctId :: Maybe Bool

Specifies whether (true) or not (false) the key identifier is distinct from the created API key value. This parameter is deprecated and should not be used.

name :: Maybe Text

The name of the ApiKey.

stageKeys :: Maybe [StageKey]

DEPRECATED FOR USAGE PLANS - Specifies stages associated with the API key.

description :: Maybe Text

The description of the ApiKey.

tags :: Maybe (HashMap Text Text)

The key-value map of strings. The valid character set is [a-zA-Z+-=._:/]. The tag key can be up to 128 characters and must not start with aws:. The tag value can be up to 256 characters.

newCreateApiKey :: CreateApiKey

Create a value of CreateApiKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:CreateApiKey', createApiKey_enabled - Specifies whether the ApiKey can be used by callers.

$sel:value:CreateApiKey', createApiKey_value - Specifies a value of the API key.

$sel:customerId:CreateApiKey', createApiKey_customerId - An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

$sel:generateDistinctId:CreateApiKey', createApiKey_generateDistinctId - Specifies whether (true) or not (false) the key identifier is distinct from the created API key value. This parameter is deprecated and should not be used.

$sel:name:CreateApiKey', createApiKey_name - The name of the ApiKey.

$sel:stageKeys:CreateApiKey', createApiKey_stageKeys - DEPRECATED FOR USAGE PLANS - Specifies stages associated with the API key.

$sel:description:CreateApiKey', createApiKey_description - The description of the ApiKey.

$sel:tags:CreateApiKey', createApiKey_tags - The key-value map of strings. The valid character set is [a-zA-Z+-=._:/]. The tag key can be up to 128 characters and must not start with aws:. The tag value can be up to 256 characters.

Request Lenses

createApiKey_enabled :: Lens' CreateApiKey (Maybe Bool)

Specifies whether the ApiKey can be used by callers.

createApiKey_value :: Lens' CreateApiKey (Maybe Text)

Specifies a value of the API key.

createApiKey_customerId :: Lens' CreateApiKey (Maybe Text)

An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

createApiKey_generateDistinctId :: Lens' CreateApiKey (Maybe Bool)

Specifies whether (true) or not (false) the key identifier is distinct from the created API key value. This parameter is deprecated and should not be used.

createApiKey_name :: Lens' CreateApiKey (Maybe Text)

The name of the ApiKey.

createApiKey_stageKeys :: Lens' CreateApiKey (Maybe [StageKey])

DEPRECATED FOR USAGE PLANS - Specifies stages associated with the API key.

createApiKey_description :: Lens' CreateApiKey (Maybe Text)

The description of the ApiKey.

createApiKey_tags :: Lens' CreateApiKey (Maybe (HashMap Text Text))

The key-value map of strings. The valid character set is [a-zA-Z+-=._:/]. The tag key can be up to 128 characters and must not start with aws:. The tag value can be up to 256 characters.

Destructuring the Response

data ApiKey

A resource that can be distributed to callers for executing Method resources that require an API key. API keys can be mapped to any Stage on any RestApi, which indicates that the callers with the API key can make requests to that stage.

Use API Keys

See: newApiKey smart constructor.

Constructors

ApiKey'

Fields

enabled :: Maybe Bool

Specifies whether the API Key can be used by callers.

value :: Maybe Text

The value of the API Key.

customerId :: Maybe Text

An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

createdDate :: Maybe POSIX

The timestamp when the API Key was created.

name :: Maybe Text

The name of the API Key.

id :: Maybe Text

The identifier of the API Key.

stageKeys :: Maybe [Text]

A list of Stage resources that are associated with the ApiKey resource.

lastUpdatedDate :: Maybe POSIX

The timestamp when the API Key was last updated.

description :: Maybe Text

The description of the API Key.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

newApiKey :: ApiKey

Create a value of ApiKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:ApiKey', apiKey_enabled - Specifies whether the API Key can be used by callers.

$sel:value:ApiKey', apiKey_value - The value of the API Key.

$sel:customerId:ApiKey', apiKey_customerId - An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

$sel:createdDate:ApiKey', apiKey_createdDate - The timestamp when the API Key was created.

$sel:name:ApiKey', apiKey_name - The name of the API Key.

$sel:id:ApiKey', apiKey_id - The identifier of the API Key.

$sel:stageKeys:ApiKey', apiKey_stageKeys - A list of Stage resources that are associated with the ApiKey resource.

$sel:lastUpdatedDate:ApiKey', apiKey_lastUpdatedDate - The timestamp when the API Key was last updated.

$sel:description:ApiKey', apiKey_description - The description of the API Key.

$sel:tags:ApiKey', apiKey_tags - The collection of tags. Each tag element is associated with a given resource.

Response Lenses

apiKey_enabled :: Lens' ApiKey (Maybe Bool)

Specifies whether the API Key can be used by callers.

apiKey_value :: Lens' ApiKey (Maybe Text)

The value of the API Key.

apiKey_customerId :: Lens' ApiKey (Maybe Text)

An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

apiKey_createdDate :: Lens' ApiKey (Maybe UTCTime)

The timestamp when the API Key was created.

apiKey_name :: Lens' ApiKey (Maybe Text)

The name of the API Key.

apiKey_id :: Lens' ApiKey (Maybe Text)

The identifier of the API Key.

apiKey_stageKeys :: Lens' ApiKey (Maybe [Text])

A list of Stage resources that are associated with the ApiKey resource.

apiKey_lastUpdatedDate :: Lens' ApiKey (Maybe UTCTime)

The timestamp when the API Key was last updated.

apiKey_description :: Lens' ApiKey (Maybe Text)

The description of the API Key.

apiKey_tags :: Lens' ApiKey (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.