Network.AWS.APIGateway.CreateModel

Description

Adds a new Model resource to an existing RestApi resource.

Creating a Request

data CreateModel

Request to add a new Model to an existing RestApi resource.

See: newCreateModel smart constructor.

Constructors

CreateModel'

Fields

schema :: Maybe Text

The schema for the model. For application/json models, this should be JSON schema draft 4 model.

description :: Maybe Text

The description of the model.

restApiId :: Text

Required

The RestApi identifier under which the Model will be created.

name :: Text

Required

The name of the model. Must be alphanumeric.

contentType :: Text

Required

The content-type for the model.

newCreateModel

Arguments

:: Text
$sel:restApiId:CreateModel'

-> Text
$sel:name:CreateModel'

-> Text
$sel:contentType:CreateModel'

-> CreateModel

Create a value of CreateModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schema:CreateModel', createModel_schema - The schema for the model. For application/json models, this should be JSON schema draft 4 model.

$sel:description:CreateModel', createModel_description - The description of the model.

$sel:restApiId:CreateModel', createModel_restApiId - [Required] The RestApi identifier under which the Model will be created.

$sel:name:CreateModel', createModel_name - [Required] The name of the model. Must be alphanumeric.

$sel:contentType:CreateModel', createModel_contentType - [Required] The content-type for the model.

Request Lenses

createModel_schema :: Lens' CreateModel (Maybe Text)

The schema for the model. For application/json models, this should be JSON schema draft 4 model.

createModel_description :: Lens' CreateModel (Maybe Text)

The description of the model.

createModel_restApiId :: Lens' CreateModel Text

Required

The RestApi identifier under which the Model will be created.

createModel_name :: Lens' CreateModel Text

Required

The name of the model. Must be alphanumeric.

createModel_contentType :: Lens' CreateModel Text

Required

The content-type for the model.

Destructuring the Response

data Model

Represents the data structure of a method's request or response payload.

A request model defines the data structure of the client-supplied request payload. A response model defines the data structure of the response payload returned by the back end. Although not required, models are useful for mapping payloads between the front end and back end.

A model is used for generating an API's SDK, validating the input request body, and creating a skeletal mapping template.

Method, MethodResponse, Models and Mappings

See: newModel smart constructor.

Constructors

Model'

Fields

schema :: Maybe Text

The schema for the model. For application/json models, this should be JSON schema draft 4 model. Do not include "\*/" characters in the description of any properties because such "\*/" characters may be interpreted as the closing marker for comments in some languages, such as Java or JavaScript, causing the installation of your API's SDK generated by API Gateway to fail.

contentType :: Maybe Text

The content-type for the model.

id :: Maybe Text

The identifier for the model resource.

name :: Maybe Text

The name of the model. Must be an alphanumeric string.

description :: Maybe Text

The description of the model.

newModel :: Model

Create a value of Model with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schema:Model', model_schema - The schema for the model. For application/json models, this should be JSON schema draft 4 model. Do not include "\*/" characters in the description of any properties because such "\*/" characters may be interpreted as the closing marker for comments in some languages, such as Java or JavaScript, causing the installation of your API's SDK generated by API Gateway to fail.

$sel:contentType:Model', model_contentType - The content-type for the model.

$sel:id:Model', model_id - The identifier for the model resource.

$sel:name:Model', model_name - The name of the model. Must be an alphanumeric string.

$sel:description:Model', model_description - The description of the model.

Response Lenses

model_schema :: Lens' Model (Maybe Text)

The schema for the model. For application/json models, this should be JSON schema draft 4 model. Do not include "\*/" characters in the description of any properties because such "\*/" characters may be interpreted as the closing marker for comments in some languages, such as Java or JavaScript, causing the installation of your API's SDK generated by API Gateway to fail.

model_contentType :: Lens' Model (Maybe Text)

The content-type for the model.

model_id :: Lens' Model (Maybe Text)

The identifier for the model resource.

model_name :: Lens' Model (Maybe Text)

The name of the model. Must be an alphanumeric string.

model_description :: Lens' Model (Maybe Text)

The description of the model.