Network.AWS.APIGateway.CreateStage

Description

Creates a new Stage resource that references a pre-existing Deployment for the API.

Creating a Request

data CreateStage

Requests API Gateway to create a Stage resource.

See: newCreateStage smart constructor.

Constructors

CreateStage'

Fields

tracingEnabled :: Maybe Bool

Specifies whether active tracing with X-ray is enabled for the Stage.

cacheClusterEnabled :: Maybe Bool

Whether cache clustering is enabled for the stage.

documentationVersion :: Maybe Text

The version of the associated API documentation.

variables :: Maybe (HashMap Text Text)

A map that defines the stage variables for the new Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

tags :: Maybe (HashMap Text Text)

The key-value map of strings. The valid character set is [a-zA-Z+-=._:/]. The tag key can be up to 128 characters and must not start with aws:. The tag value can be up to 256 characters.

description :: Maybe Text

The description of the Stage resource.

cacheClusterSize :: Maybe CacheClusterSize

The stage's cache cluster size.

canarySettings :: Maybe CanarySettings

The canary deployment settings of this stage.

restApiId :: Text

Required

The string identifier of the associated RestApi.

stageName :: Text

Required

The name for the Stage resource. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

deploymentId :: Text

Required

The identifier of the Deployment resource for the Stage resource.

newCreateStage

Arguments

:: Text
$sel:restApiId:CreateStage'

-> Text
$sel:stageName:CreateStage'

-> Text
$sel:deploymentId:CreateStage'

-> CreateStage

Create a value of CreateStage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tracingEnabled:CreateStage', createStage_tracingEnabled - Specifies whether active tracing with X-ray is enabled for the Stage.

$sel:cacheClusterEnabled:CreateStage', createStage_cacheClusterEnabled - Whether cache clustering is enabled for the stage.

$sel:documentationVersion:CreateStage', createStage_documentationVersion - The version of the associated API documentation.

$sel:variables:CreateStage', createStage_variables - A map that defines the stage variables for the new Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

$sel:tags:CreateStage', createStage_tags - The key-value map of strings. The valid character set is [a-zA-Z+-=._:/]. The tag key can be up to 128 characters and must not start with aws:. The tag value can be up to 256 characters.

$sel:description:CreateStage', createStage_description - The description of the Stage resource.

$sel:cacheClusterSize:CreateStage', createStage_cacheClusterSize - The stage's cache cluster size.

$sel:canarySettings:CreateStage', createStage_canarySettings - The canary deployment settings of this stage.

$sel:restApiId:CreateStage', createStage_restApiId - [Required] The string identifier of the associated RestApi.

$sel:stageName:CreateStage', createStage_stageName - [Required] The name for the Stage resource. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

$sel:deploymentId:CreateStage', createStage_deploymentId - [Required] The identifier of the Deployment resource for the Stage resource.

Request Lenses

createStage_tracingEnabled :: Lens' CreateStage (Maybe Bool)

Specifies whether active tracing with X-ray is enabled for the Stage.

createStage_cacheClusterEnabled :: Lens' CreateStage (Maybe Bool)

Whether cache clustering is enabled for the stage.

createStage_documentationVersion :: Lens' CreateStage (Maybe Text)

The version of the associated API documentation.

createStage_variables :: Lens' CreateStage (Maybe (HashMap Text Text))

A map that defines the stage variables for the new Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

createStage_tags :: Lens' CreateStage (Maybe (HashMap Text Text))

The key-value map of strings. The valid character set is [a-zA-Z+-=._:/]. The tag key can be up to 128 characters and must not start with aws:. The tag value can be up to 256 characters.

createStage_description :: Lens' CreateStage (Maybe Text)

The description of the Stage resource.

createStage_cacheClusterSize :: Lens' CreateStage (Maybe CacheClusterSize)

The stage's cache cluster size.

createStage_canarySettings :: Lens' CreateStage (Maybe CanarySettings)

The canary deployment settings of this stage.

createStage_restApiId :: Lens' CreateStage Text

Required

The string identifier of the associated RestApi.

createStage_stageName :: Lens' CreateStage Text

Required

The name for the Stage resource. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

createStage_deploymentId :: Lens' CreateStage Text

Required

The identifier of the Deployment resource for the Stage resource.

Destructuring the Response

data Stage

Represents a unique identifier for a version of a deployed RestApi that is callable by users.

Deploy an API

See: newStage smart constructor.

Constructors

Stage'

Fields

createdDate :: Maybe POSIX

The timestamp when the stage was created.

deploymentId :: Maybe Text

The identifier of the Deployment that the stage points to.

tracingEnabled :: Maybe Bool

Specifies whether active tracing with X-ray is enabled for the Stage.

lastUpdatedDate :: Maybe POSIX

The timestamp when the stage last updated.

cacheClusterEnabled :: Maybe Bool

Specifies whether a cache cluster is enabled for the stage.

stageName :: Maybe Text

The name of the stage is the first path segment in the Uniform Resource Identifier (URI) of a call to API Gateway. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

webAclArn :: Maybe Text

The ARN of the WebAcl associated with the Stage.

documentationVersion :: Maybe Text

The version of the associated API documentation.

accessLogSettings :: Maybe AccessLogSettings

Settings for logging access in this stage.

variables :: Maybe (HashMap Text Text)

A map that defines the stage variables for a Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

clientCertificateId :: Maybe Text

The identifier of a client certificate for an API stage.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

description :: Maybe Text

The stage's description.

cacheClusterSize :: Maybe CacheClusterSize

The size of the cache cluster for the stage, if enabled.

canarySettings :: Maybe CanarySettings

Settings for the canary deployment in this stage.

cacheClusterStatus :: Maybe CacheClusterStatus

The status of the cache cluster for the stage, if enabled.

methodSettings :: Maybe (HashMap Text MethodSetting)

A map that defines the method settings for a Stage resource. Keys (designated as /{method_setting_key below) are method paths defined as {resource_path}/{http_method} for an individual method override, or /\*/\* for overriding all methods in the stage.

newStage :: Stage

Create a value of Stage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:Stage', stage_createdDate - The timestamp when the stage was created.

$sel:deploymentId:Stage', stage_deploymentId - The identifier of the Deployment that the stage points to.

$sel:tracingEnabled:Stage', stage_tracingEnabled - Specifies whether active tracing with X-ray is enabled for the Stage.

$sel:lastUpdatedDate:Stage', stage_lastUpdatedDate - The timestamp when the stage last updated.

$sel:cacheClusterEnabled:Stage', stage_cacheClusterEnabled - Specifies whether a cache cluster is enabled for the stage.

$sel:stageName:Stage', stage_stageName - The name of the stage is the first path segment in the Uniform Resource Identifier (URI) of a call to API Gateway. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

$sel:webAclArn:Stage', stage_webAclArn - The ARN of the WebAcl associated with the Stage.

$sel:documentationVersion:Stage', stage_documentationVersion - The version of the associated API documentation.

$sel:accessLogSettings:Stage', stage_accessLogSettings - Settings for logging access in this stage.

$sel:variables:Stage', stage_variables - A map that defines the stage variables for a Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

$sel:clientCertificateId:Stage', stage_clientCertificateId - The identifier of a client certificate for an API stage.

$sel:tags:Stage', stage_tags - The collection of tags. Each tag element is associated with a given resource.

$sel:description:Stage', stage_description - The stage's description.

$sel:cacheClusterSize:Stage', stage_cacheClusterSize - The size of the cache cluster for the stage, if enabled.

$sel:canarySettings:Stage', stage_canarySettings - Settings for the canary deployment in this stage.

$sel:cacheClusterStatus:Stage', stage_cacheClusterStatus - The status of the cache cluster for the stage, if enabled.

$sel:methodSettings:Stage', stage_methodSettings - A map that defines the method settings for a Stage resource. Keys (designated as /{method_setting_key below) are method paths defined as {resource_path}/{http_method} for an individual method override, or /\*/\* for overriding all methods in the stage.

Response Lenses

stage_createdDate :: Lens' Stage (Maybe UTCTime)

The timestamp when the stage was created.

stage_deploymentId :: Lens' Stage (Maybe Text)

The identifier of the Deployment that the stage points to.

stage_tracingEnabled :: Lens' Stage (Maybe Bool)

Specifies whether active tracing with X-ray is enabled for the Stage.

stage_lastUpdatedDate :: Lens' Stage (Maybe UTCTime)

The timestamp when the stage last updated.

stage_cacheClusterEnabled :: Lens' Stage (Maybe Bool)

Specifies whether a cache cluster is enabled for the stage.

stage_stageName :: Lens' Stage (Maybe Text)

The name of the stage is the first path segment in the Uniform Resource Identifier (URI) of a call to API Gateway. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

stage_webAclArn :: Lens' Stage (Maybe Text)

The ARN of the WebAcl associated with the Stage.

stage_documentationVersion :: Lens' Stage (Maybe Text)

The version of the associated API documentation.

stage_accessLogSettings :: Lens' Stage (Maybe AccessLogSettings)

Settings for logging access in this stage.

stage_variables :: Lens' Stage (Maybe (HashMap Text Text))

A map that defines the stage variables for a Stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

stage_clientCertificateId :: Lens' Stage (Maybe Text)

The identifier of a client certificate for an API stage.

stage_tags :: Lens' Stage (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.

stage_description :: Lens' Stage (Maybe Text)

The stage's description.

stage_cacheClusterSize :: Lens' Stage (Maybe CacheClusterSize)

The size of the cache cluster for the stage, if enabled.

stage_canarySettings :: Lens' Stage (Maybe CanarySettings)

Settings for the canary deployment in this stage.

stage_cacheClusterStatus :: Lens' Stage (Maybe CacheClusterStatus)

The status of the cache cluster for the stage, if enabled.

stage_methodSettings :: Lens' Stage (Maybe (HashMap Text MethodSetting))

A map that defines the method settings for a Stage resource. Keys (designated as /{method_setting_key below) are method paths defined as {resource_path}/{http_method} for an individual method override, or /\*/\* for overriding all methods in the stage.