Network.AWS.APIGateway.CreateUsagePlanKey

Description

Creates a usage plan key for adding an existing API key to a usage plan.

Creating a Request

data CreateUsagePlanKey

The POST request to create a usage plan key for adding an existing API key to a usage plan.

See: newCreateUsagePlanKey smart constructor.

Constructors

CreateUsagePlanKey'

Fields

usagePlanId :: Text

Required

The Id of the UsagePlan resource representing the usage plan containing the to-be-created UsagePlanKey resource representing a plan customer.

keyId :: Text

Required

The identifier of a UsagePlanKey resource for a plan customer.

keyType :: Text

Required

The type of a UsagePlanKey resource for a plan customer.

newCreateUsagePlanKey

Arguments

:: Text
$sel:usagePlanId:CreateUsagePlanKey'

-> Text
$sel:keyId:CreateUsagePlanKey'

-> Text
$sel:keyType:CreateUsagePlanKey'

-> CreateUsagePlanKey

Create a value of CreateUsagePlanKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:usagePlanId:CreateUsagePlanKey', createUsagePlanKey_usagePlanId - [Required] The Id of the UsagePlan resource representing the usage plan containing the to-be-created UsagePlanKey resource representing a plan customer.

$sel:keyId:CreateUsagePlanKey', createUsagePlanKey_keyId - [Required] The identifier of a UsagePlanKey resource for a plan customer.

$sel:keyType:CreateUsagePlanKey', createUsagePlanKey_keyType - [Required] The type of a UsagePlanKey resource for a plan customer.

Request Lenses

createUsagePlanKey_usagePlanId :: Lens' CreateUsagePlanKey Text

Required

The Id of the UsagePlan resource representing the usage plan containing the to-be-created UsagePlanKey resource representing a plan customer.

createUsagePlanKey_keyId :: Lens' CreateUsagePlanKey Text

Required

The identifier of a UsagePlanKey resource for a plan customer.

createUsagePlanKey_keyType :: Lens' CreateUsagePlanKey Text

Required

The type of a UsagePlanKey resource for a plan customer.

Destructuring the Response

data UsagePlanKey

Represents a usage plan key to identify a plan customer.

To associate an API stage with a selected API key in a usage plan, you must create a UsagePlanKey resource to represent the selected ApiKey.

"

Create and Use Usage Plans

See: newUsagePlanKey smart constructor.

Constructors

UsagePlanKey'

Fields

id :: Maybe Text

The Id of a usage plan key.

name :: Maybe Text

The name of a usage plan key.

value :: Maybe Text

The value of a usage plan key.

type' :: Maybe Text

The type of a usage plan key. Currently, the valid key type is API_KEY.

newUsagePlanKey :: UsagePlanKey

Create a value of UsagePlanKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:UsagePlanKey', usagePlanKey_id - The Id of a usage plan key.

$sel:name:UsagePlanKey', usagePlanKey_name - The name of a usage plan key.

$sel:value:UsagePlanKey', usagePlanKey_value - The value of a usage plan key.

$sel:type':UsagePlanKey', usagePlanKey_type - The type of a usage plan key. Currently, the valid key type is API_KEY.

Response Lenses

usagePlanKey_id :: Lens' UsagePlanKey (Maybe Text)

The Id of a usage plan key.

usagePlanKey_name :: Lens' UsagePlanKey (Maybe Text)

The name of a usage plan key.

usagePlanKey_value :: Lens' UsagePlanKey (Maybe Text)

The value of a usage plan key.

usagePlanKey_type :: Lens' UsagePlanKey (Maybe Text)

The type of a usage plan key. Currently, the valid key type is API_KEY.