Network.AWS.APIGateway.DeleteGatewayResponse

Description

Clears any customization of a GatewayResponse of a specified response type on the given RestApi and resets it with the default settings.

Creating a Request

data DeleteGatewayResponse

Clears any customization of a GatewayResponse of a specified response type on the given RestApi and resets it with the default settings.

See: newDeleteGatewayResponse smart constructor.

Constructors

DeleteGatewayResponse'

Fields

restApiId :: Text

Required

The string identifier of the associated RestApi.

responseType :: GatewayResponseType

Required

The response type of the associated GatewayResponse.

newDeleteGatewayResponse

Arguments

:: Text
$sel:restApiId:DeleteGatewayResponse'

-> GatewayResponseType
$sel:responseType:DeleteGatewayResponse'

-> DeleteGatewayResponse

Create a value of DeleteGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:restApiId:DeleteGatewayResponse', deleteGatewayResponse_restApiId - [Required] The string identifier of the associated RestApi.

$sel:responseType:DeleteGatewayResponse', deleteGatewayResponse_responseType - [Required]

The response type of the associated GatewayResponse.

Request Lenses

deleteGatewayResponse_restApiId :: Lens' DeleteGatewayResponse Text

Required

The string identifier of the associated RestApi.

deleteGatewayResponse_responseType :: Lens' DeleteGatewayResponse GatewayResponseType

Required

The response type of the associated GatewayResponse.

Destructuring the Response

data DeleteGatewayResponseResponse

See: newDeleteGatewayResponseResponse smart constructor.

Constructors

DeleteGatewayResponseResponse'

newDeleteGatewayResponseResponse :: DeleteGatewayResponseResponse

Create a value of DeleteGatewayResponseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.