Network.AWS.APIGateway.GetApiKeys

Description

Gets information about the current ApiKeys resource.

This operation returns paginated results.

Creating a Request

data GetApiKeys

A request to get information about the current ApiKeys resource.

See: newGetApiKeys smart constructor.

Constructors

GetApiKeys'

Fields

includeValues :: Maybe Bool

A boolean flag to specify whether (true) or not (false) the result contains key values.

customerId :: Maybe Text

The identifier of a customer in AWS Marketplace or an external system, such as a developer portal.

nameQuery :: Maybe Text

The name of queried API keys.

limit :: Maybe Int

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

position :: Maybe Text

The current pagination position in the paged result set.

newGetApiKeys :: GetApiKeys

Create a value of GetApiKeys with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeValues:GetApiKeys', getApiKeys_includeValues - A boolean flag to specify whether (true) or not (false) the result contains key values.

$sel:customerId:GetApiKeys', getApiKeys_customerId - The identifier of a customer in AWS Marketplace or an external system, such as a developer portal.

$sel:nameQuery:GetApiKeys', getApiKeys_nameQuery - The name of queried API keys.

$sel:limit:GetApiKeys', getApiKeys_limit - The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

$sel:position:GetApiKeys', getApiKeys_position - The current pagination position in the paged result set.

Request Lenses

getApiKeys_includeValues :: Lens' GetApiKeys (Maybe Bool)

A boolean flag to specify whether (true) or not (false) the result contains key values.

getApiKeys_customerId :: Lens' GetApiKeys (Maybe Text)

The identifier of a customer in AWS Marketplace or an external system, such as a developer portal.

getApiKeys_nameQuery :: Lens' GetApiKeys (Maybe Text)

The name of queried API keys.

getApiKeys_limit :: Lens' GetApiKeys (Maybe Int)

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

getApiKeys_position :: Lens' GetApiKeys (Maybe Text)

The current pagination position in the paged result set.

Destructuring the Response

data GetApiKeysResponse

Represents a collection of API keys as represented by an ApiKeys resource.

Use API Keys

See: newGetApiKeysResponse smart constructor.

Constructors

GetApiKeysResponse'

Fields

items :: Maybe [ApiKey]

The current page of elements from this collection.

warnings :: Maybe [Text]

A list of warning messages logged during the import of API keys when the failOnWarnings option is set to true.

position :: Maybe Text

httpStatus :: Int

The response's http status code.

newGetApiKeysResponse

Arguments

:: Int
$sel:httpStatus:GetApiKeysResponse'

-> GetApiKeysResponse

Create a value of GetApiKeysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:GetApiKeysResponse', getApiKeysResponse_items - The current page of elements from this collection.

$sel:warnings:GetApiKeysResponse', getApiKeysResponse_warnings - A list of warning messages logged during the import of API keys when the failOnWarnings option is set to true.

$sel:position:GetApiKeys', getApiKeysResponse_position - Undocumented member.

$sel:httpStatus:GetApiKeysResponse', getApiKeysResponse_httpStatus - The response's http status code.

Response Lenses

getApiKeysResponse_items :: Lens' GetApiKeysResponse (Maybe [ApiKey])

The current page of elements from this collection.

getApiKeysResponse_warnings :: Lens' GetApiKeysResponse (Maybe [Text])

A list of warning messages logged during the import of API keys when the failOnWarnings option is set to true.

getApiKeysResponse_position :: Lens' GetApiKeysResponse (Maybe Text)

Undocumented member.

getApiKeysResponse_httpStatus :: Lens' GetApiKeysResponse Int

The response's http status code.