Network.AWS.APIGateway.GetBasePathMapping

Description

Describe a BasePathMapping resource.

Creating a Request

data GetBasePathMapping

Request to describe a BasePathMapping resource.

See: newGetBasePathMapping smart constructor.

Constructors

GetBasePathMapping'

Fields

domainName :: Text

Required

The domain name of the BasePathMapping resource to be described.

basePath :: Text

Required

The base path name that callers of the API must provide as part of the URL after the domain name. This value must be unique for all of the mappings across a single API. Specify '(none)' if you do not want callers to specify any base path name after the domain name.

newGetBasePathMapping

Arguments

:: Text
$sel:domainName:GetBasePathMapping'

-> Text
$sel:basePath:GetBasePathMapping'

-> GetBasePathMapping

Create a value of GetBasePathMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:GetBasePathMapping', getBasePathMapping_domainName - [Required] The domain name of the BasePathMapping resource to be described.

$sel:basePath:GetBasePathMapping', getBasePathMapping_basePath - [Required] The base path name that callers of the API must provide as part of the URL after the domain name. This value must be unique for all of the mappings across a single API. Specify '(none)' if you do not want callers to specify any base path name after the domain name.

Request Lenses

getBasePathMapping_domainName :: Lens' GetBasePathMapping Text

Required

The domain name of the BasePathMapping resource to be described.

getBasePathMapping_basePath :: Lens' GetBasePathMapping Text

Required

The base path name that callers of the API must provide as part of the URL after the domain name. This value must be unique for all of the mappings across a single API. Specify '(none)' if you do not want callers to specify any base path name after the domain name.

Destructuring the Response

data BasePathMapping

Represents the base path that callers of the API must provide as part of the URL after the domain name.

A custom domain name plus a BasePathMapping specification identifies a deployed RestApi in a given stage of the owner Account.

Use Custom Domain Names

See: newBasePathMapping smart constructor.

Constructors

BasePathMapping'

Fields

basePath :: Maybe Text

The base path name that callers of the API must provide as part of the URL after the domain name.

stage :: Maybe Text

The name of the associated stage.

restApiId :: Maybe Text

The string identifier of the associated RestApi.

newBasePathMapping :: BasePathMapping

Create a value of BasePathMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:basePath:BasePathMapping', basePathMapping_basePath - The base path name that callers of the API must provide as part of the URL after the domain name.

$sel:stage:BasePathMapping', basePathMapping_stage - The name of the associated stage.

$sel:restApiId:BasePathMapping', basePathMapping_restApiId - The string identifier of the associated RestApi.

Response Lenses

basePathMapping_basePath :: Lens' BasePathMapping (Maybe Text)

The base path name that callers of the API must provide as part of the URL after the domain name.

basePathMapping_stage :: Lens' BasePathMapping (Maybe Text)

The name of the associated stage.

basePathMapping_restApiId :: Lens' BasePathMapping (Maybe Text)

The string identifier of the associated RestApi.