Network.AWS.APIGateway.GetDeployment

Description

Gets information about a Deployment resource.

Creating a Request

data GetDeployment

Requests API Gateway to get information about a Deployment resource.

See: newGetDeployment smart constructor.

Constructors

GetDeployment'

Fields

embed :: Maybe [Text]

A query parameter to retrieve the specified embedded resources of the returned Deployment resource in the response. In a REST API call, this embed parameter value is a list of comma-separated strings, as in GET /restapis/{restapi_id}/deployments/{deployment_id}?embed=var1,var2. The SDK and other platform-dependent libraries might use a different format for the list. Currently, this request supports only retrieval of the embedded API summary this way. Hence, the parameter value must be a single-valued list containing only the "apisummary" string. For example, GET /restapis/{restapi_id}/deployments/{deployment_id}?embed=apisummary.

restApiId :: Text

Required

The string identifier of the associated RestApi.

deploymentId :: Text

Required

The identifier of the Deployment resource to get information about.

newGetDeployment

Arguments

:: Text
$sel:restApiId:GetDeployment'

-> Text
$sel:deploymentId:GetDeployment'

-> GetDeployment

Create a value of GetDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:embed:GetDeployment', getDeployment_embed - A query parameter to retrieve the specified embedded resources of the returned Deployment resource in the response. In a REST API call, this embed parameter value is a list of comma-separated strings, as in GET /restapis/{restapi_id}/deployments/{deployment_id}?embed=var1,var2. The SDK and other platform-dependent libraries might use a different format for the list. Currently, this request supports only retrieval of the embedded API summary this way. Hence, the parameter value must be a single-valued list containing only the "apisummary" string. For example, GET /restapis/{restapi_id}/deployments/{deployment_id}?embed=apisummary.

$sel:restApiId:GetDeployment', getDeployment_restApiId - [Required] The string identifier of the associated RestApi.

$sel:deploymentId:GetDeployment', getDeployment_deploymentId - [Required] The identifier of the Deployment resource to get information about.

Request Lenses

getDeployment_embed :: Lens' GetDeployment (Maybe [Text])

A query parameter to retrieve the specified embedded resources of the returned Deployment resource in the response. In a REST API call, this embed parameter value is a list of comma-separated strings, as in GET /restapis/{restapi_id}/deployments/{deployment_id}?embed=var1,var2. The SDK and other platform-dependent libraries might use a different format for the list. Currently, this request supports only retrieval of the embedded API summary this way. Hence, the parameter value must be a single-valued list containing only the "apisummary" string. For example, GET /restapis/{restapi_id}/deployments/{deployment_id}?embed=apisummary.

getDeployment_restApiId :: Lens' GetDeployment Text

Required

The string identifier of the associated RestApi.

getDeployment_deploymentId :: Lens' GetDeployment Text

Required

The identifier of the Deployment resource to get information about.

Destructuring the Response

data Deployment

An immutable representation of a RestApi resource that can be called by users using Stages. A deployment must be associated with a Stage for it to be callable over the Internet.

To create a deployment, call POST on the Deployments resource of a RestApi. To view, update, or delete a deployment, call GET, PATCH, or DELETE on the specified deployment resource (/restapis/{restapi_id}/deployments/{deployment_id}).

RestApi, Deployments, Stage, AWS CLI, AWS SDKs

See: newDeployment smart constructor.

Constructors

Deployment'

Fields

apiSummary :: Maybe (HashMap Text (HashMap Text MethodSnapshot))

A summary of the RestApi at the date and time that the deployment resource was created.

createdDate :: Maybe POSIX

The date and time that the deployment resource was created.

id :: Maybe Text

The identifier for the deployment resource.

description :: Maybe Text

The description for the deployment resource.

newDeployment :: Deployment

Create a value of Deployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiSummary:Deployment', deployment_apiSummary - A summary of the RestApi at the date and time that the deployment resource was created.

$sel:createdDate:Deployment', deployment_createdDate - The date and time that the deployment resource was created.

$sel:id:Deployment', deployment_id - The identifier for the deployment resource.

$sel:description:Deployment', deployment_description - The description for the deployment resource.

Response Lenses

deployment_apiSummary :: Lens' Deployment (Maybe (HashMap Text (HashMap Text MethodSnapshot)))

A summary of the RestApi at the date and time that the deployment resource was created.

deployment_createdDate :: Lens' Deployment (Maybe UTCTime)

The date and time that the deployment resource was created.

deployment_id :: Lens' Deployment (Maybe Text)

The identifier for the deployment resource.

deployment_description :: Lens' Deployment (Maybe Text)

The description for the deployment resource.