Network.AWS.APIGateway.GetDocumentationPart

Description

- | Undocumented operation.

Creating a Request

data GetDocumentationPart

Gets a specified documentation part of a given API.

See: newGetDocumentationPart smart constructor.

Constructors

GetDocumentationPart'

Fields

restApiId :: Text

Required

The string identifier of the associated RestApi.

documentationPartId :: Text

Required

The string identifier of the associated RestApi.

newGetDocumentationPart

Arguments

:: Text
$sel:restApiId:GetDocumentationPart'

-> Text
$sel:documentationPartId:GetDocumentationPart'

-> GetDocumentationPart

Create a value of GetDocumentationPart with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:restApiId:GetDocumentationPart', getDocumentationPart_restApiId - [Required] The string identifier of the associated RestApi.

$sel:documentationPartId:GetDocumentationPart', getDocumentationPart_documentationPartId - [Required] The string identifier of the associated RestApi.

Request Lenses

getDocumentationPart_restApiId :: Lens' GetDocumentationPart Text

Required

The string identifier of the associated RestApi.

getDocumentationPart_documentationPartId :: Lens' GetDocumentationPart Text

Required

The string identifier of the associated RestApi.

Destructuring the Response

data DocumentationPart

A documentation part for a targeted API entity.

A documentation part consists of a content map (properties) and a target (location). The target specifies an API entity to which the documentation content applies. The supported API entity types are API, AUTHORIZER, MODEL, RESOURCE, METHOD, PATH_PARAMETER, QUERY_PARAMETER, REQUEST_HEADER, REQUEST_BODY, RESPONSE, RESPONSE_HEADER, and RESPONSE_BODY. Valid location fields depend on the API entity type. All valid fields are not required.

The content map is a JSON string of API-specific key-value pairs. Although an API can use any shape for the content map, only the OpenAPI-compliant documentation fields will be injected into the associated API entity definition in the exported OpenAPI definition file.

Documenting an API, DocumentationParts

See: newDocumentationPart smart constructor.

Constructors

DocumentationPart'

Fields

location :: Maybe DocumentationPartLocation

The location of the API entity to which the documentation applies. Valid fields depend on the targeted API entity type. All the valid location fields are not required. If not explicitly specified, a valid location field is treated as a wildcard and associated documentation content may be inherited by matching entities, unless overridden.

id :: Maybe Text

The DocumentationPart identifier, generated by API Gateway when the DocumentationPart is created.

properties :: Maybe Text

A content map of API-specific key-value pairs describing the targeted API entity. The map must be encoded as a JSON string, e.g., "{ \"description\": \"The API does ...\" }". Only OpenAPI-compliant documentation-related fields from the properties map are exported and, hence, published as part of the API entity definitions, while the original documentation parts are exported in a OpenAPI extension of x-amazon-apigateway-documentation.

newDocumentationPart :: DocumentationPart

Create a value of DocumentationPart with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:location:DocumentationPart', documentationPart_location - The location of the API entity to which the documentation applies. Valid fields depend on the targeted API entity type. All the valid location fields are not required. If not explicitly specified, a valid location field is treated as a wildcard and associated documentation content may be inherited by matching entities, unless overridden.

$sel:id:DocumentationPart', documentationPart_id - The DocumentationPart identifier, generated by API Gateway when the DocumentationPart is created.

$sel:properties:DocumentationPart', documentationPart_properties - A content map of API-specific key-value pairs describing the targeted API entity. The map must be encoded as a JSON string, e.g., "{ \"description\": \"The API does ...\" }". Only OpenAPI-compliant documentation-related fields from the properties map are exported and, hence, published as part of the API entity definitions, while the original documentation parts are exported in a OpenAPI extension of x-amazon-apigateway-documentation.

Response Lenses

documentationPart_location :: Lens' DocumentationPart (Maybe DocumentationPartLocation)

The location of the API entity to which the documentation applies. Valid fields depend on the targeted API entity type. All the valid location fields are not required. If not explicitly specified, a valid location field is treated as a wildcard and associated documentation content may be inherited by matching entities, unless overridden.

documentationPart_id :: Lens' DocumentationPart (Maybe Text)

The DocumentationPart identifier, generated by API Gateway when the DocumentationPart is created.

documentationPart_properties :: Lens' DocumentationPart (Maybe Text)

A content map of API-specific key-value pairs describing the targeted API entity. The map must be encoded as a JSON string, e.g., "{ \"description\": \"The API does ...\" }". Only OpenAPI-compliant documentation-related fields from the properties map are exported and, hence, published as part of the API entity definitions, while the original documentation parts are exported in a OpenAPI extension of x-amazon-apigateway-documentation.