Network.AWS.APIGateway.GetDocumentationVersion

Description

- | Undocumented operation.

Creating a Request

data GetDocumentationVersion

Gets a documentation snapshot of an API.

See: newGetDocumentationVersion smart constructor.

Constructors

GetDocumentationVersion'

Fields

restApiId :: Text

Required

The string identifier of the associated RestApi.

documentationVersion :: Text

Required

The version identifier of the to-be-retrieved documentation snapshot.

newGetDocumentationVersion

Arguments

:: Text
$sel:restApiId:GetDocumentationVersion'

-> Text
$sel:documentationVersion:GetDocumentationVersion'

-> GetDocumentationVersion

Create a value of GetDocumentationVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:restApiId:GetDocumentationVersion', getDocumentationVersion_restApiId - [Required] The string identifier of the associated RestApi.

$sel:documentationVersion:GetDocumentationVersion', getDocumentationVersion_documentationVersion - [Required] The version identifier of the to-be-retrieved documentation snapshot.

Request Lenses

getDocumentationVersion_restApiId :: Lens' GetDocumentationVersion Text

Required

The string identifier of the associated RestApi.

getDocumentationVersion_documentationVersion :: Lens' GetDocumentationVersion Text

Required

The version identifier of the to-be-retrieved documentation snapshot.

Destructuring the Response

data DocumentationVersion

A snapshot of the documentation of an API.

Publishing API documentation involves creating a documentation version associated with an API stage and exporting the versioned documentation to an external (e.g., OpenAPI) file.

Documenting an API, DocumentationPart, DocumentationVersions

See: newDocumentationVersion smart constructor.

Constructors

DocumentationVersion'

Fields

createdDate :: Maybe POSIX

The date when the API documentation snapshot is created.

version :: Maybe Text

The version identifier of the API documentation snapshot.

description :: Maybe Text

The description of the API documentation snapshot.

newDocumentationVersion :: DocumentationVersion

Create a value of DocumentationVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:DocumentationVersion', documentationVersion_createdDate - The date when the API documentation snapshot is created.

$sel:version:DocumentationVersion', documentationVersion_version - The version identifier of the API documentation snapshot.

$sel:description:DocumentationVersion', documentationVersion_description - The description of the API documentation snapshot.

Response Lenses

documentationVersion_createdDate :: Lens' DocumentationVersion (Maybe UTCTime)

The date when the API documentation snapshot is created.

documentationVersion_version :: Lens' DocumentationVersion (Maybe Text)

The version identifier of the API documentation snapshot.

documentationVersion_description :: Lens' DocumentationVersion (Maybe Text)

The description of the API documentation snapshot.