Network.AWS.APIGateway.GetDomainName

Description

Represents a domain name that is contained in a simpler, more intuitive URL that can be called.

Creating a Request

data GetDomainName

Request to get the name of a DomainName resource.

See: newGetDomainName smart constructor.

Constructors

GetDomainName'

Fields

domainName :: Text

Required

The name of the DomainName resource.

newGetDomainName

Arguments

:: Text
$sel:domainName:GetDomainName'

-> GetDomainName

Create a value of GetDomainName with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:GetDomainName', getDomainName_domainName - [Required] The name of the DomainName resource.

Request Lenses

getDomainName_domainName :: Lens' GetDomainName Text

Required

The name of the DomainName resource.

Destructuring the Response

data DomainName

Represents a custom domain name as a user-friendly host name of an API (RestApi).

When you deploy an API, API Gateway creates a default host name for the API. This default API host name is of the {restapi-id}.execute-api.{region}.amazonaws.com format. With the default host name, you can access the API's root resource with the URL of https://{restapi-id}.execute-api.{region}.amazonaws.com/{stage}/. When you set up a custom domain name of apis.example.com for this API, you can then access the same resource using the URL of the https://apis.examples.com/myApi, where myApi is the base path mapping (BasePathMapping) of your API under the custom domain name.

Set a Custom Host Name for an API

See: newDomainName smart constructor.

Constructors

DomainName'

Fields

regionalHostedZoneId :: Maybe Text

The region-specific Amazon Route 53 Hosted Zone ID of the regional endpoint. For more information, see Set up a Regional Custom Domain Name and AWS Regions and Endpoints for API Gateway.

certificateName :: Maybe Text

The name of the certificate that will be used by edge-optimized endpoint for this domain name.

ownershipVerificationCertificateArn :: Maybe Text

The ARN of the public certificate issued by ACM to validate ownership of your custom domain. Only required when configuring mutual TLS and using an ACM imported or private CA certificate ARN as the regionalCertificateArn.

regionalCertificateArn :: Maybe Text

The reference to an AWS-managed certificate that will be used for validating the regional domain name. AWS Certificate Manager is the only supported source.

certificateArn :: Maybe Text

The reference to an AWS-managed certificate that will be used by edge-optimized endpoint for this domain name. AWS Certificate Manager is the only supported source.

distributionHostedZoneId :: Maybe Text

The region-agnostic Amazon Route 53 Hosted Zone ID of the edge-optimized endpoint. The valid value is Z2FDTNDATAQYW2 for all the regions. For more information, see Set up a Regional Custom Domain Name and AWS Regions and Endpoints for API Gateway.

securityPolicy :: Maybe SecurityPolicy

The Transport Layer Security (TLS) version + cipher suite for this DomainName. The valid values are TLS_1_0 and TLS_1_2.

domainName :: Maybe Text

The custom domain name as an API host name, for example, my-api.example.com.

mutualTlsAuthentication :: Maybe MutualTlsAuthentication

The mutual TLS authentication configuration for a custom domain name. If specified, API Gateway performs two-way authentication between the client and the server. Clients must present a trusted certificate to access your API.

regionalCertificateName :: Maybe Text

The name of the certificate that will be used for validating the regional domain name.

regionalDomainName :: Maybe Text

The domain name associated with the regional endpoint for this custom domain name. You set up this association by adding a DNS record that points the custom domain name to this regional domain name. The regional domain name is returned by API Gateway when you create a regional endpoint.

certificateUploadDate :: Maybe POSIX

The timestamp when the certificate that was used by edge-optimized endpoint for this domain name was uploaded.

distributionDomainName :: Maybe Text

The domain name of the Amazon CloudFront distribution associated with this custom domain name for an edge-optimized endpoint. You set up this association when adding a DNS record pointing the custom domain name to this distribution name. For more information about CloudFront distributions, see the Amazon CloudFront documentation.

domainNameStatusMessage :: Maybe Text

An optional text message containing detailed information about status of the DomainName migration.

endpointConfiguration :: Maybe EndpointConfiguration

The endpoint configuration of this DomainName showing the endpoint types of the domain name.

domainNameStatus :: Maybe DomainNameStatus

The status of the DomainName migration. The valid values are AVAILABLE, UPDATING, PENDING_CERTIFICATE_REIMPORT, and PENDING_OWNERSHIP_VERIFICATION. If the status is UPDATING, the domain cannot be modified further until the existing operation is complete. If it is AVAILABLE, the domain can be updated.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

newDomainName :: DomainName

Create a value of DomainName with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regionalHostedZoneId:DomainName', domainName_regionalHostedZoneId - The region-specific Amazon Route 53 Hosted Zone ID of the regional endpoint. For more information, see Set up a Regional Custom Domain Name and AWS Regions and Endpoints for API Gateway.

$sel:certificateName:DomainName', domainName_certificateName - The name of the certificate that will be used by edge-optimized endpoint for this domain name.

$sel:ownershipVerificationCertificateArn:DomainName', domainName_ownershipVerificationCertificateArn - The ARN of the public certificate issued by ACM to validate ownership of your custom domain. Only required when configuring mutual TLS and using an ACM imported or private CA certificate ARN as the regionalCertificateArn.

$sel:regionalCertificateArn:DomainName', domainName_regionalCertificateArn - The reference to an AWS-managed certificate that will be used for validating the regional domain name. AWS Certificate Manager is the only supported source.

$sel:certificateArn:DomainName', domainName_certificateArn - The reference to an AWS-managed certificate that will be used by edge-optimized endpoint for this domain name. AWS Certificate Manager is the only supported source.

$sel:distributionHostedZoneId:DomainName', domainName_distributionHostedZoneId - The region-agnostic Amazon Route 53 Hosted Zone ID of the edge-optimized endpoint. The valid value is Z2FDTNDATAQYW2 for all the regions. For more information, see Set up a Regional Custom Domain Name and AWS Regions and Endpoints for API Gateway.

$sel:securityPolicy:DomainName', domainName_securityPolicy - The Transport Layer Security (TLS) version + cipher suite for this DomainName. The valid values are TLS_1_0 and TLS_1_2.

$sel:domainName:DomainName', domainName_domainName - The custom domain name as an API host name, for example, my-api.example.com.

$sel:mutualTlsAuthentication:DomainName', domainName_mutualTlsAuthentication - The mutual TLS authentication configuration for a custom domain name. If specified, API Gateway performs two-way authentication between the client and the server. Clients must present a trusted certificate to access your API.

$sel:regionalCertificateName:DomainName', domainName_regionalCertificateName - The name of the certificate that will be used for validating the regional domain name.

$sel:regionalDomainName:DomainName', domainName_regionalDomainName - The domain name associated with the regional endpoint for this custom domain name. You set up this association by adding a DNS record that points the custom domain name to this regional domain name. The regional domain name is returned by API Gateway when you create a regional endpoint.

$sel:certificateUploadDate:DomainName', domainName_certificateUploadDate - The timestamp when the certificate that was used by edge-optimized endpoint for this domain name was uploaded.

$sel:distributionDomainName:DomainName', domainName_distributionDomainName - The domain name of the Amazon CloudFront distribution associated with this custom domain name for an edge-optimized endpoint. You set up this association when adding a DNS record pointing the custom domain name to this distribution name. For more information about CloudFront distributions, see the Amazon CloudFront documentation.

$sel:domainNameStatusMessage:DomainName', domainName_domainNameStatusMessage - An optional text message containing detailed information about status of the DomainName migration.

$sel:endpointConfiguration:DomainName', domainName_endpointConfiguration - The endpoint configuration of this DomainName showing the endpoint types of the domain name.

$sel:domainNameStatus:DomainName', domainName_domainNameStatus - The status of the DomainName migration. The valid values are AVAILABLE, UPDATING, PENDING_CERTIFICATE_REIMPORT, and PENDING_OWNERSHIP_VERIFICATION. If the status is UPDATING, the domain cannot be modified further until the existing operation is complete. If it is AVAILABLE, the domain can be updated.

$sel:tags:DomainName', domainName_tags - The collection of tags. Each tag element is associated with a given resource.

Response Lenses

domainName_regionalHostedZoneId :: Lens' DomainName (Maybe Text)

The region-specific Amazon Route 53 Hosted Zone ID of the regional endpoint. For more information, see Set up a Regional Custom Domain Name and AWS Regions and Endpoints for API Gateway.

domainName_certificateName :: Lens' DomainName (Maybe Text)

The name of the certificate that will be used by edge-optimized endpoint for this domain name.

domainName_ownershipVerificationCertificateArn :: Lens' DomainName (Maybe Text)

The ARN of the public certificate issued by ACM to validate ownership of your custom domain. Only required when configuring mutual TLS and using an ACM imported or private CA certificate ARN as the regionalCertificateArn.

domainName_regionalCertificateArn :: Lens' DomainName (Maybe Text)

The reference to an AWS-managed certificate that will be used for validating the regional domain name. AWS Certificate Manager is the only supported source.

domainName_certificateArn :: Lens' DomainName (Maybe Text)

The reference to an AWS-managed certificate that will be used by edge-optimized endpoint for this domain name. AWS Certificate Manager is the only supported source.

domainName_distributionHostedZoneId :: Lens' DomainName (Maybe Text)

The region-agnostic Amazon Route 53 Hosted Zone ID of the edge-optimized endpoint. The valid value is Z2FDTNDATAQYW2 for all the regions. For more information, see Set up a Regional Custom Domain Name and AWS Regions and Endpoints for API Gateway.

domainName_securityPolicy :: Lens' DomainName (Maybe SecurityPolicy)

The Transport Layer Security (TLS) version + cipher suite for this DomainName. The valid values are TLS_1_0 and TLS_1_2.

domainName_domainName :: Lens' DomainName (Maybe Text)

The custom domain name as an API host name, for example, my-api.example.com.

domainName_mutualTlsAuthentication :: Lens' DomainName (Maybe MutualTlsAuthentication)

The mutual TLS authentication configuration for a custom domain name. If specified, API Gateway performs two-way authentication between the client and the server. Clients must present a trusted certificate to access your API.

domainName_regionalCertificateName :: Lens' DomainName (Maybe Text)

The name of the certificate that will be used for validating the regional domain name.

domainName_regionalDomainName :: Lens' DomainName (Maybe Text)

The domain name associated with the regional endpoint for this custom domain name. You set up this association by adding a DNS record that points the custom domain name to this regional domain name. The regional domain name is returned by API Gateway when you create a regional endpoint.

domainName_certificateUploadDate :: Lens' DomainName (Maybe UTCTime)

The timestamp when the certificate that was used by edge-optimized endpoint for this domain name was uploaded.

domainName_distributionDomainName :: Lens' DomainName (Maybe Text)

The domain name of the Amazon CloudFront distribution associated with this custom domain name for an edge-optimized endpoint. You set up this association when adding a DNS record pointing the custom domain name to this distribution name. For more information about CloudFront distributions, see the Amazon CloudFront documentation.

domainName_domainNameStatusMessage :: Lens' DomainName (Maybe Text)

An optional text message containing detailed information about status of the DomainName migration.

domainName_endpointConfiguration :: Lens' DomainName (Maybe EndpointConfiguration)

The endpoint configuration of this DomainName showing the endpoint types of the domain name.

domainName_domainNameStatus :: Lens' DomainName (Maybe DomainNameStatus)

The status of the DomainName migration. The valid values are AVAILABLE, UPDATING, PENDING_CERTIFICATE_REIMPORT, and PENDING_OWNERSHIP_VERIFICATION. If the status is UPDATING, the domain cannot be modified further until the existing operation is complete. If it is AVAILABLE, the domain can be updated.

domainName_tags :: Lens' DomainName (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.