Network.AWS.APIGateway.GetExport

Description

Exports a deployed version of a RestApi in a specified format.

Creating a Request

data GetExport

Request a new export of a RestApi for a particular Stage.

See: newGetExport smart constructor.

Constructors

GetExport'

Fields

parameters :: Maybe (HashMap Text Text)

A key-value map of query string parameters that specify properties of the export, depending on the requested exportType. For exportType oas30 and swagger, any combination of the following parameters are supported: extensions='integrations' or extensions='apigateway' will export the API with x-amazon-apigateway-integration extensions. extensions='authorizers' will export the API with x-amazon-apigateway-authorizer extensions. postman will export the API with Postman extensions, allowing for import to the Postman tool

accepts :: Maybe Text

The content-type of the export, for example application/json. Currently application/json and application/yaml are supported for exportType ofoas30 and swagger. This should be specified in the Accept header for direct API requests.

restApiId :: Text

Required

The string identifier of the associated RestApi.

stageName :: Text

Required

The name of the Stage that will be exported.

exportType :: Text

Required

The type of export. Acceptable values are 'oas30' for OpenAPI 3.0.x and 'swagger' for Swagger/OpenAPI 2.0.

newGetExport

Arguments

:: Text
$sel:restApiId:GetExport'

-> Text
$sel:stageName:GetExport'

-> Text
$sel:exportType:GetExport'

-> GetExport

Create a value of GetExport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:GetExport', getExport_parameters - A key-value map of query string parameters that specify properties of the export, depending on the requested exportType. For exportType oas30 and swagger, any combination of the following parameters are supported: extensions='integrations' or extensions='apigateway' will export the API with x-amazon-apigateway-integration extensions. extensions='authorizers' will export the API with x-amazon-apigateway-authorizer extensions. postman will export the API with Postman extensions, allowing for import to the Postman tool

$sel:accepts:GetExport', getExport_accepts - The content-type of the export, for example application/json. Currently application/json and application/yaml are supported for exportType ofoas30 and swagger. This should be specified in the Accept header for direct API requests.

$sel:restApiId:GetExport', getExport_restApiId - [Required] The string identifier of the associated RestApi.

$sel:stageName:GetExport', getExport_stageName - [Required] The name of the Stage that will be exported.

$sel:exportType:GetExport', getExport_exportType - [Required] The type of export. Acceptable values are 'oas30' for OpenAPI 3.0.x and 'swagger' for Swagger/OpenAPI 2.0.

Request Lenses

getExport_parameters :: Lens' GetExport (Maybe (HashMap Text Text))

A key-value map of query string parameters that specify properties of the export, depending on the requested exportType. For exportType oas30 and swagger, any combination of the following parameters are supported: extensions='integrations' or extensions='apigateway' will export the API with x-amazon-apigateway-integration extensions. extensions='authorizers' will export the API with x-amazon-apigateway-authorizer extensions. postman will export the API with Postman extensions, allowing for import to the Postman tool

getExport_accepts :: Lens' GetExport (Maybe Text)

The content-type of the export, for example application/json. Currently application/json and application/yaml are supported for exportType ofoas30 and swagger. This should be specified in the Accept header for direct API requests.

getExport_restApiId :: Lens' GetExport Text

Required

The string identifier of the associated RestApi.

getExport_stageName :: Lens' GetExport Text

Required

The name of the Stage that will be exported.

getExport_exportType :: Lens' GetExport Text

Required

The type of export. Acceptable values are 'oas30' for OpenAPI 3.0.x and 'swagger' for Swagger/OpenAPI 2.0.

Destructuring the Response

data GetExportResponse

The binary blob response to GetExport, which contains the generated SDK.

See: newGetExportResponse smart constructor.

Constructors

GetExportResponse'

Fields

body :: Maybe ByteString

The binary blob response to GetExport, which contains the export.

contentDisposition :: Maybe Text

The content-disposition header value in the HTTP response.

contentType :: Maybe Text

The content-type header value in the HTTP response. This will correspond to a valid 'accept' type in the request.

httpStatus :: Int

The response's http status code.

newGetExportResponse

Arguments

:: Int
$sel:httpStatus:GetExportResponse'

-> GetExportResponse

Create a value of GetExportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:body:GetExportResponse', getExportResponse_body - The binary blob response to GetExport, which contains the export.

$sel:contentDisposition:GetExportResponse', getExportResponse_contentDisposition - The content-disposition header value in the HTTP response.

$sel:contentType:GetExportResponse', getExportResponse_contentType - The content-type header value in the HTTP response. This will correspond to a valid 'accept' type in the request.

$sel:httpStatus:GetExportResponse', getExportResponse_httpStatus - The response's http status code.

Response Lenses

getExportResponse_body :: Lens' GetExportResponse (Maybe ByteString)

The binary blob response to GetExport, which contains the export.

getExportResponse_contentDisposition :: Lens' GetExportResponse (Maybe Text)

The content-disposition header value in the HTTP response.

getExportResponse_contentType :: Lens' GetExportResponse (Maybe Text)

The content-type header value in the HTTP response. This will correspond to a valid 'accept' type in the request.

getExportResponse_httpStatus :: Lens' GetExportResponse Int

The response's http status code.