Network.AWS.APIGateway.GetMethodResponse

Description

Describes a MethodResponse resource.

Creating a Request

data GetMethodResponse

Request to describe a MethodResponse resource.

See: newGetMethodResponse smart constructor.

Constructors

GetMethodResponse'

Fields

restApiId :: Text

Required

The string identifier of the associated RestApi.

resourceId :: Text

Required

The Resource identifier for the MethodResponse resource.

httpMethod :: Text

Required

The HTTP verb of the Method resource.

statusCode :: Text

Required

The status code for the MethodResponse resource.

newGetMethodResponse

Arguments

:: Text
$sel:restApiId:GetMethodResponse'

-> Text
$sel:resourceId:GetMethodResponse'

-> Text
$sel:httpMethod:GetMethodResponse'

-> Text
$sel:statusCode:GetMethodResponse'

-> GetMethodResponse

Create a value of GetMethodResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:restApiId:GetMethodResponse', getMethodResponse_restApiId - [Required] The string identifier of the associated RestApi.

$sel:resourceId:GetMethodResponse', getMethodResponse_resourceId - [Required] The Resource identifier for the MethodResponse resource.

$sel:httpMethod:GetMethodResponse', getMethodResponse_httpMethod - [Required] The HTTP verb of the Method resource.

$sel:statusCode:GetMethodResponse', getMethodResponse_statusCode - [Required] The status code for the MethodResponse resource.

Request Lenses

getMethodResponse_restApiId :: Lens' GetMethodResponse Text

Required

The string identifier of the associated RestApi.

getMethodResponse_resourceId :: Lens' GetMethodResponse Text

Required

The Resource identifier for the MethodResponse resource.

getMethodResponse_httpMethod :: Lens' GetMethodResponse Text

Required

The HTTP verb of the Method resource.

getMethodResponse_statusCode :: Lens' GetMethodResponse Text

Required

The status code for the MethodResponse resource.

Destructuring the Response

data MethodResponse

Represents a method response of a given HTTP status code returned to the client. The method response is passed from the back end through the associated integration response that can be transformed using a mapping template.

Example: A MethodResponse instance of an API

Request

The example request retrieves a MethodResponse of the 200 status code.

GET /restapis/fugvjdxtri/resources/3kzxbg5sa2/methods/GET/responses/200 HTTP/1.1 Content-Type: application/json Host: apigateway.us-east-1.amazonaws.com X-Amz-Date: 20160603T222952Z Authorization: AWS4-HMAC-SHA256 Credential={access_key_ID}/20160603/us-east-1/apigateway/aws4_request, SignedHeaders=content-type;host;x-amz-date, Signature={sig4_hash}

Response

The successful response returns 200 OK status and a payload as follows:

{ "_links": { "curies": { "href": "https://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-method-response-{rel}.html", "name": "methodresponse", "templated": true }, "self": { "href": "/restapis/fugvjdxtri/resources/3kzxbg5sa2/methods/GET/responses/200", "title": "200" }, "methodresponse:delete": { "href": "/restapis/fugvjdxtri/resources/3kzxbg5sa2/methods/GET/responses/200" }, "methodresponse:update": { "href": "/restapis/fugvjdxtri/resources/3kzxbg5sa2/methods/GET/responses/200" } }, "responseModels": { "application/json": "Empty" }, "responseParameters": { "method.response.header.Content-Type": false }, "statusCode": "200" }

Method, IntegrationResponse, Integration Creating an API

See: newMethodResponse smart constructor.

Constructors

MethodResponse'

Fields

responseModels :: Maybe (HashMap Text Text)

Specifies the Model resources used for the response's content-type. Response models are represented as a key/value map, with a content-type as the key and a Model name as the value.

statusCode :: Maybe Text

The method response's status code.

responseParameters :: Maybe (HashMap Text Bool)

A key-value map specifying required or optional response parameters that API Gateway can send back to the caller. A key defines a method response header and the value specifies whether the associated method response header is required or not. The expression of the key must match the pattern method.response.header.{name}, where name is a valid and unique header name. API Gateway passes certain integration response data to the method response headers specified here according to the mapping you prescribe in the API's IntegrationResponse. The integration response data that can be mapped include an integration response header expressed in integration.response.header.{name}, a static value enclosed within a pair of single quotes (e.g., 'application/json'), or a JSON expression from the back-end response payload in the form of integration.response.body.{JSON-expression}, where JSON-expression is a valid JSON expression without the $ prefix.)

newMethodResponse :: MethodResponse

Create a value of MethodResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:responseModels:MethodResponse', methodResponse_responseModels - Specifies the Model resources used for the response's content-type. Response models are represented as a key/value map, with a content-type as the key and a Model name as the value.

$sel:statusCode:MethodResponse', methodResponse_statusCode - The method response's status code.

$sel:responseParameters:MethodResponse', methodResponse_responseParameters - A key-value map specifying required or optional response parameters that API Gateway can send back to the caller. A key defines a method response header and the value specifies whether the associated method response header is required or not. The expression of the key must match the pattern method.response.header.{name}, where name is a valid and unique header name. API Gateway passes certain integration response data to the method response headers specified here according to the mapping you prescribe in the API's IntegrationResponse. The integration response data that can be mapped include an integration response header expressed in integration.response.header.{name}, a static value enclosed within a pair of single quotes (e.g., 'application/json'), or a JSON expression from the back-end response payload in the form of integration.response.body.{JSON-expression}, where JSON-expression is a valid JSON expression without the $ prefix.)

Response Lenses

methodResponse_responseModels :: Lens' MethodResponse (Maybe (HashMap Text Text))

Specifies the Model resources used for the response's content-type. Response models are represented as a key/value map, with a content-type as the key and a Model name as the value.

methodResponse_statusCode :: Lens' MethodResponse (Maybe Text)

The method response's status code.

methodResponse_responseParameters :: Lens' MethodResponse (Maybe (HashMap Text Bool))

A key-value map specifying required or optional response parameters that API Gateway can send back to the caller. A key defines a method response header and the value specifies whether the associated method response header is required or not. The expression of the key must match the pattern method.response.header.{name}, where name is a valid and unique header name. API Gateway passes certain integration response data to the method response headers specified here according to the mapping you prescribe in the API's IntegrationResponse. The integration response data that can be mapped include an integration response header expressed in integration.response.header.{name}, a static value enclosed within a pair of single quotes (e.g., 'application/json'), or a JSON expression from the back-end response payload in the form of integration.response.body.{JSON-expression}, where JSON-expression is a valid JSON expression without the $ prefix.)