Network.AWS.APIGateway.GetModelTemplate

Description

Generates a sample mapping template that can be used to transform a payload into the structure of a model.

Creating a Request

data GetModelTemplate

Request to generate a sample mapping template used to transform the payload.

See: newGetModelTemplate smart constructor.

Constructors

GetModelTemplate'

Fields

restApiId :: Text

Required

The string identifier of the associated RestApi.

modelName :: Text

Required

The name of the model for which to generate a template.

newGetModelTemplate

Arguments

:: Text
$sel:restApiId:GetModelTemplate'

-> Text
$sel:modelName:GetModelTemplate'

-> GetModelTemplate

Create a value of GetModelTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:restApiId:GetModelTemplate', getModelTemplate_restApiId - [Required] The string identifier of the associated RestApi.

$sel:modelName:GetModelTemplate', getModelTemplate_modelName - [Required] The name of the model for which to generate a template.

Request Lenses

getModelTemplate_restApiId :: Lens' GetModelTemplate Text

Required

The string identifier of the associated RestApi.

getModelTemplate_modelName :: Lens' GetModelTemplate Text

Required

The name of the model for which to generate a template.

Destructuring the Response

data GetModelTemplateResponse

Represents a mapping template used to transform a payload.

Mapping Templates

See: newGetModelTemplateResponse smart constructor.

Constructors

GetModelTemplateResponse'

Fields

value :: Maybe Text

The Apache Velocity Template Language (VTL) template content used for the template resource.

httpStatus :: Int

The response's http status code.

newGetModelTemplateResponse

Arguments

:: Int
$sel:httpStatus:GetModelTemplateResponse'

-> GetModelTemplateResponse

Create a value of GetModelTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:GetModelTemplateResponse', getModelTemplateResponse_value - The Apache Velocity Template Language (VTL) template content used for the template resource.

$sel:httpStatus:GetModelTemplateResponse', getModelTemplateResponse_httpStatus - The response's http status code.

Response Lenses

getModelTemplateResponse_value :: Lens' GetModelTemplateResponse (Maybe Text)

The Apache Velocity Template Language (VTL) template content used for the template resource.

getModelTemplateResponse_httpStatus :: Lens' GetModelTemplateResponse Int

The response's http status code.