Network.AWS.APIGateway.GetRequestValidators

Description

Gets the RequestValidators collection of a given RestApi.

This operation returns paginated results.

Creating a Request

data GetRequestValidators

Gets the RequestValidators collection of a given RestApi.

See: newGetRequestValidators smart constructor.

Constructors

GetRequestValidators'

Fields

position :: Maybe Text

The current pagination position in the paged result set.

limit :: Maybe Int

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

restApiId :: Text

Required

The string identifier of the associated RestApi.

newGetRequestValidators

Arguments

:: Text
$sel:restApiId:GetRequestValidators'

-> GetRequestValidators

Create a value of GetRequestValidators with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:position:GetRequestValidators', getRequestValidators_position - The current pagination position in the paged result set.

$sel:limit:GetRequestValidators', getRequestValidators_limit - The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

$sel:restApiId:GetRequestValidators', getRequestValidators_restApiId - [Required] The string identifier of the associated RestApi.

Request Lenses

getRequestValidators_position :: Lens' GetRequestValidators (Maybe Text)

The current pagination position in the paged result set.

getRequestValidators_limit :: Lens' GetRequestValidators (Maybe Int)

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

getRequestValidators_restApiId :: Lens' GetRequestValidators Text

Required

The string identifier of the associated RestApi.

Destructuring the Response

data GetRequestValidatorsResponse

A collection of RequestValidator resources of a given RestApi.

In OpenAPI, the RequestValidators of an API is defined by the x-amazon-apigateway-request-validators extension.

Enable Basic Request Validation in API Gateway

See: newGetRequestValidatorsResponse smart constructor.

Constructors

GetRequestValidatorsResponse'

Fields

items :: Maybe [RequestValidator]

The current page of elements from this collection.

position :: Maybe Text

httpStatus :: Int

The response's http status code.

newGetRequestValidatorsResponse

Arguments

:: Int
$sel:httpStatus:GetRequestValidatorsResponse'

-> GetRequestValidatorsResponse

Create a value of GetRequestValidatorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:GetRequestValidatorsResponse', getRequestValidatorsResponse_items - The current page of elements from this collection.

$sel:position:GetRequestValidators', getRequestValidatorsResponse_position - Undocumented member.

$sel:httpStatus:GetRequestValidatorsResponse', getRequestValidatorsResponse_httpStatus - The response's http status code.

Response Lenses

getRequestValidatorsResponse_items :: Lens' GetRequestValidatorsResponse (Maybe [RequestValidator])

The current page of elements from this collection.

getRequestValidatorsResponse_position :: Lens' GetRequestValidatorsResponse (Maybe Text)

Undocumented member.

getRequestValidatorsResponse_httpStatus :: Lens' GetRequestValidatorsResponse Int

The response's http status code.