Network.AWS.APIGateway.GetSdkType

Description

- | Undocumented operation.

Creating a Request

data GetSdkType

Get an SdkType instance.

See: newGetSdkType smart constructor.

Constructors

GetSdkType'

Fields

id :: Text

Required

The identifier of the queried SdkType instance.

newGetSdkType

Arguments

:: Text
$sel:id:GetSdkType'

-> GetSdkType

Create a value of GetSdkType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetSdkType', getSdkType_id - [Required] The identifier of the queried SdkType instance.

Request Lenses

getSdkType_id :: Lens' GetSdkType Text

Required

The identifier of the queried SdkType instance.

Destructuring the Response

data SdkType

A type of SDK that API Gateway can generate.

See: newSdkType smart constructor.

Constructors

SdkType'

Fields

friendlyName :: Maybe Text

The user-friendly name of an SdkType instance.

id :: Maybe Text

The identifier of an SdkType instance.

configurationProperties :: Maybe [SdkConfigurationProperty]

A list of configuration properties of an SdkType.

description :: Maybe Text

The description of an SdkType.

newSdkType :: SdkType

Create a value of SdkType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:friendlyName:SdkType', sdkType_friendlyName - The user-friendly name of an SdkType instance.

$sel:id:SdkType', sdkType_id - The identifier of an SdkType instance.

$sel:configurationProperties:SdkType', sdkType_configurationProperties - A list of configuration properties of an SdkType.

$sel:description:SdkType', sdkType_description - The description of an SdkType.

Response Lenses

sdkType_friendlyName :: Lens' SdkType (Maybe Text)

The user-friendly name of an SdkType instance.

sdkType_id :: Lens' SdkType (Maybe Text)

The identifier of an SdkType instance.

sdkType_configurationProperties :: Lens' SdkType (Maybe [SdkConfigurationProperty])

A list of configuration properties of an SdkType.

sdkType_description :: Lens' SdkType (Maybe Text)

The description of an SdkType.