Network.AWS.APIGateway.GetTags

Description

Gets the Tags collection for a given resource.

Creating a Request

data GetTags

Gets the Tags collection for a given resource.

See: newGetTags smart constructor.

Constructors

GetTags'

Fields

position :: Maybe Text

(Not currently supported) The current pagination position in the paged result set.

limit :: Maybe Int

(Not currently supported) The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

resourceArn :: Text

Required

The ARN of a resource that can be tagged.

newGetTags

Arguments

:: Text
$sel:resourceArn:GetTags'

-> GetTags

Create a value of GetTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:position:GetTags', getTags_position - (Not currently supported) The current pagination position in the paged result set.

$sel:limit:GetTags', getTags_limit - (Not currently supported) The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

$sel:resourceArn:GetTags', getTags_resourceArn - [Required] The ARN of a resource that can be tagged.

Request Lenses

getTags_position :: Lens' GetTags (Maybe Text)

(Not currently supported) The current pagination position in the paged result set.

getTags_limit :: Lens' GetTags (Maybe Int)

(Not currently supported) The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

getTags_resourceArn :: Lens' GetTags Text

Required

The ARN of a resource that can be tagged.

Destructuring the Response

data GetTagsResponse

The collection of tags. Each tag element is associated with a given resource.

See: newGetTagsResponse smart constructor.

Constructors

GetTagsResponse'

Fields

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

httpStatus :: Int

The response's http status code.

newGetTagsResponse

Arguments

:: Int
$sel:httpStatus:GetTagsResponse'

-> GetTagsResponse

Create a value of GetTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:GetTagsResponse', getTagsResponse_tags - The collection of tags. Each tag element is associated with a given resource.

$sel:httpStatus:GetTagsResponse', getTagsResponse_httpStatus - The response's http status code.

Response Lenses

getTagsResponse_tags :: Lens' GetTagsResponse (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.

getTagsResponse_httpStatus :: Lens' GetTagsResponse Int

The response's http status code.