Network.AWS.APIGateway.GetUsage

Description

Gets the usage data of a usage plan in a specified time interval.

This operation returns paginated results.

Creating a Request

data GetUsage

The GET request to get the usage data of a usage plan in a specified time interval.

See: newGetUsage smart constructor.

Constructors

GetUsage'

Fields

position :: Maybe Text

The current pagination position in the paged result set.

limit :: Maybe Int

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

keyId :: Maybe Text

The Id of the API key associated with the resultant usage data.

usagePlanId :: Text

Required

The Id of the usage plan associated with the usage data.

startDate :: Text

Required

The starting date (e.g., 2016-01-01) of the usage data.

endDate :: Text

Required

The ending date (e.g., 2016-12-31) of the usage data.

newGetUsage

Arguments

:: Text
$sel:usagePlanId:GetUsage'

-> Text
$sel:startDate:GetUsage'

-> Text
$sel:endDate:GetUsage'

-> GetUsage

Create a value of GetUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:position:GetUsage', getUsage_position - The current pagination position in the paged result set.

$sel:limit:GetUsage', getUsage_limit - The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

$sel:keyId:GetUsage', getUsage_keyId - The Id of the API key associated with the resultant usage data.

$sel:usagePlanId:GetUsage', getUsage_usagePlanId - [Required] The Id of the usage plan associated with the usage data.

$sel:startDate:GetUsage', getUsage_startDate - [Required] The starting date (e.g., 2016-01-01) of the usage data.

$sel:endDate:GetUsage', getUsage_endDate - [Required] The ending date (e.g., 2016-12-31) of the usage data.

Request Lenses

getUsage_position :: Lens' GetUsage (Maybe Text)

The current pagination position in the paged result set.

getUsage_limit :: Lens' GetUsage (Maybe Int)

The maximum number of returned results per page. The default value is 25 and the maximum value is 500.

getUsage_keyId :: Lens' GetUsage (Maybe Text)

The Id of the API key associated with the resultant usage data.

getUsage_usagePlanId :: Lens' GetUsage Text

Required

The Id of the usage plan associated with the usage data.

getUsage_startDate :: Lens' GetUsage Text

Required

The starting date (e.g., 2016-01-01) of the usage data.

getUsage_endDate :: Lens' GetUsage Text

Required

The ending date (e.g., 2016-12-31) of the usage data.

Destructuring the Response

data Usage

Represents the usage data of a usage plan.

Create and Use Usage Plans, Manage Usage in a Usage Plan

See: newUsage smart constructor.

Constructors

Usage'

Fields

startDate :: Maybe Text

The starting date of the usage data.

items :: Maybe (HashMap Text [[Integer]])

The usage data, as daily logs of used and remaining quotas, over the specified time interval indexed over the API keys in a usage plan. For example, {..., "values" : { "{api_key}" : [ [0, 100], [10, 90], [100, 10]]}, where {api_key} stands for an API key value and the daily log entry is of the format [used quota, remaining quota].

position :: Maybe Text

endDate :: Maybe Text

The ending date of the usage data.

usagePlanId :: Maybe Text

The plan Id associated with this usage data.

newUsage :: Usage

Create a value of Usage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startDate:Usage', usage_startDate - The starting date of the usage data.

$sel:items:Usage', usage_items - The usage data, as daily logs of used and remaining quotas, over the specified time interval indexed over the API keys in a usage plan. For example, {..., "values" : { "{api_key}" : [ [0, 100], [10, 90], [100, 10]]}, where {api_key} stands for an API key value and the daily log entry is of the format [used quota, remaining quota].

$sel:position:Usage', usage_position - Undocumented member.

$sel:endDate:Usage', usage_endDate - The ending date of the usage data.

$sel:usagePlanId:Usage', usage_usagePlanId - The plan Id associated with this usage data.

Response Lenses

usage_startDate :: Lens' Usage (Maybe Text)

The starting date of the usage data.

usage_items :: Lens' Usage (Maybe (HashMap Text [[Integer]]))

The usage data, as daily logs of used and remaining quotas, over the specified time interval indexed over the API keys in a usage plan. For example, {..., "values" : { "{api_key}" : [ [0, 100], [10, 90], [100, 10]]}, where {api_key} stands for an API key value and the daily log entry is of the format [used quota, remaining quota].

usage_position :: Lens' Usage (Maybe Text)

Undocumented member.

usage_endDate :: Lens' Usage (Maybe Text)

The ending date of the usage data.

usage_usagePlanId :: Lens' Usage (Maybe Text)

The plan Id associated with this usage data.