Network.AWS.APIGateway.ImportApiKeys

Description

Import API keys from an external source, such as a CSV-formatted file.

Creating a Request

data ImportApiKeys

The POST request to import API keys from an external source, such as a CSV-formatted file.

See: newImportApiKeys smart constructor.

Constructors

ImportApiKeys'

Fields

failOnWarnings :: Maybe Bool

A query parameter to indicate whether to rollback ApiKey importation (true) or not (false) when error is encountered.

body :: ByteString

The payload of the POST request to import API keys. For the payload format, see API Key File Format.

format :: ApiKeysFormat

A query parameter to specify the input format to imported API keys. Currently, only the csv format is supported.

newImportApiKeys

Arguments

:: ByteString
$sel:body:ImportApiKeys'

-> ApiKeysFormat
$sel:format:ImportApiKeys'

-> ImportApiKeys

Create a value of ImportApiKeys with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failOnWarnings:ImportApiKeys', importApiKeys_failOnWarnings - A query parameter to indicate whether to rollback ApiKey importation (true) or not (false) when error is encountered.

$sel:body:ImportApiKeys', importApiKeys_body - The payload of the POST request to import API keys. For the payload format, see API Key File Format.

$sel:format:ImportApiKeys', importApiKeys_format - A query parameter to specify the input format to imported API keys. Currently, only the csv format is supported.

Request Lenses

importApiKeys_failOnWarnings :: Lens' ImportApiKeys (Maybe Bool)

A query parameter to indicate whether to rollback ApiKey importation (true) or not (false) when error is encountered.

importApiKeys_body :: Lens' ImportApiKeys ByteString

The payload of the POST request to import API keys. For the payload format, see API Key File Format.

importApiKeys_format :: Lens' ImportApiKeys ApiKeysFormat

A query parameter to specify the input format to imported API keys. Currently, only the csv format is supported.

Destructuring the Response

data ImportApiKeysResponse

The identifier of an ApiKey used in a UsagePlan.

See: newImportApiKeysResponse smart constructor.

Constructors

ImportApiKeysResponse'

Fields

warnings :: Maybe [Text]

A list of warning messages.

ids :: Maybe [Text]

A list of all the ApiKey identifiers.

httpStatus :: Int

The response's http status code.

newImportApiKeysResponse

Arguments

:: Int
$sel:httpStatus:ImportApiKeysResponse'

-> ImportApiKeysResponse

Create a value of ImportApiKeysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:warnings:ImportApiKeysResponse', importApiKeysResponse_warnings - A list of warning messages.

$sel:ids:ImportApiKeysResponse', importApiKeysResponse_ids - A list of all the ApiKey identifiers.

$sel:httpStatus:ImportApiKeysResponse', importApiKeysResponse_httpStatus - The response's http status code.

Response Lenses

importApiKeysResponse_warnings :: Lens' ImportApiKeysResponse (Maybe [Text])

A list of warning messages.

importApiKeysResponse_ids :: Lens' ImportApiKeysResponse (Maybe [Text])

A list of all the ApiKey identifiers.

importApiKeysResponse_httpStatus :: Lens' ImportApiKeysResponse Int

The response's http status code.