Network.AWS.APIGateway.ImportDocumentationParts

Description

- | Undocumented operation.

Creating a Request

data ImportDocumentationParts

Import documentation parts from an external (e.g., OpenAPI) definition file.

See: newImportDocumentationParts smart constructor.

Constructors

ImportDocumentationParts'

Fields

mode :: Maybe PutMode

A query parameter to indicate whether to overwrite (OVERWRITE) any existing DocumentationParts definition or to merge (MERGE) the new definition into the existing one. The default value is MERGE.

failOnWarnings :: Maybe Bool

A query parameter to specify whether to rollback the documentation importation (true) or not (false) when a warning is encountered. The default value is false.

restApiId :: Text

Required

The string identifier of the associated RestApi.

body :: ByteString

Required

Raw byte array representing the to-be-imported documentation parts. To import from an OpenAPI file, this is a JSON object.

newImportDocumentationParts

Arguments

:: Text
$sel:restApiId:ImportDocumentationParts'

-> ByteString
$sel:body:ImportDocumentationParts'

-> ImportDocumentationParts

Create a value of ImportDocumentationParts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mode:ImportDocumentationParts', importDocumentationParts_mode - A query parameter to indicate whether to overwrite (OVERWRITE) any existing DocumentationParts definition or to merge (MERGE) the new definition into the existing one. The default value is MERGE.

$sel:failOnWarnings:ImportDocumentationParts', importDocumentationParts_failOnWarnings - A query parameter to specify whether to rollback the documentation importation (true) or not (false) when a warning is encountered. The default value is false.

$sel:restApiId:ImportDocumentationParts', importDocumentationParts_restApiId - [Required] The string identifier of the associated RestApi.

$sel:body:ImportDocumentationParts', importDocumentationParts_body - [Required] Raw byte array representing the to-be-imported documentation parts. To import from an OpenAPI file, this is a JSON object.

Request Lenses

importDocumentationParts_mode :: Lens' ImportDocumentationParts (Maybe PutMode)

A query parameter to indicate whether to overwrite (OVERWRITE) any existing DocumentationParts definition or to merge (MERGE) the new definition into the existing one. The default value is MERGE.

importDocumentationParts_failOnWarnings :: Lens' ImportDocumentationParts (Maybe Bool)

A query parameter to specify whether to rollback the documentation importation (true) or not (false) when a warning is encountered. The default value is false.

importDocumentationParts_restApiId :: Lens' ImportDocumentationParts Text

Required

The string identifier of the associated RestApi.

importDocumentationParts_body :: Lens' ImportDocumentationParts ByteString

Required

Raw byte array representing the to-be-imported documentation parts. To import from an OpenAPI file, this is a JSON object.

Destructuring the Response

data ImportDocumentationPartsResponse

A collection of the imported DocumentationPart identifiers.

This is used to return the result when documentation parts in an external (e.g., OpenAPI) file are imported into API Gateway

Documenting an API, documentationpart:import, DocumentationPart

See: newImportDocumentationPartsResponse smart constructor.

Constructors

ImportDocumentationPartsResponse'

Fields

warnings :: Maybe [Text]

A list of warning messages reported during import of documentation parts.

ids :: Maybe [Text]

A list of the returned documentation part identifiers.

httpStatus :: Int

The response's http status code.

newImportDocumentationPartsResponse

Arguments

:: Int
$sel:httpStatus:ImportDocumentationPartsResponse'

-> ImportDocumentationPartsResponse

Create a value of ImportDocumentationPartsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:warnings:ImportDocumentationPartsResponse', importDocumentationPartsResponse_warnings - A list of warning messages reported during import of documentation parts.

$sel:ids:ImportDocumentationPartsResponse', importDocumentationPartsResponse_ids - A list of the returned documentation part identifiers.

$sel:httpStatus:ImportDocumentationPartsResponse', importDocumentationPartsResponse_httpStatus - The response's http status code.

Response Lenses

importDocumentationPartsResponse_warnings :: Lens' ImportDocumentationPartsResponse (Maybe [Text])

A list of warning messages reported during import of documentation parts.

importDocumentationPartsResponse_ids :: Lens' ImportDocumentationPartsResponse (Maybe [Text])

A list of the returned documentation part identifiers.

importDocumentationPartsResponse_httpStatus :: Lens' ImportDocumentationPartsResponse Int

The response's http status code.