Network.AWS.APIGateway.PutIntegrationResponse

Description

Represents a put integration.

Creating a Request

data PutIntegrationResponse

Represents a put integration response request.

See: newPutIntegrationResponse smart constructor.

Constructors

PutIntegrationResponse'

Fields

contentHandling :: Maybe ContentHandlingStrategy

Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the method response without modification.

responseTemplates :: Maybe (HashMap Text Text)

Specifies a put integration response's templates.

responseParameters :: Maybe (HashMap Text Text)

A key-value map specifying response parameters that are passed to the method response from the back end. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name must be a valid and unique response header name and JSON-expression a valid JSON expression without the $ prefix.

selectionPattern :: Maybe Text

Specifies the selection pattern of a put integration response.

restApiId :: Text

Required

The string identifier of the associated RestApi.

resourceId :: Text

Required

Specifies a put integration response request's resource identifier.

httpMethod :: Text

Required

Specifies a put integration response request's HTTP method.

statusCode :: Text

Required

Specifies the status code that is used to map the integration response to an existing MethodResponse.

newPutIntegrationResponse

Arguments

:: Text
$sel:restApiId:PutIntegrationResponse'

-> Text
$sel:resourceId:PutIntegrationResponse'

-> Text
$sel:httpMethod:PutIntegrationResponse'

-> Text
$sel:statusCode:PutIntegrationResponse'

-> PutIntegrationResponse

Create a value of PutIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentHandling:PutIntegrationResponse', putIntegrationResponse_contentHandling - Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the method response without modification.

$sel:responseTemplates:PutIntegrationResponse', putIntegrationResponse_responseTemplates - Specifies a put integration response's templates.

$sel:responseParameters:PutIntegrationResponse', putIntegrationResponse_responseParameters - A key-value map specifying response parameters that are passed to the method response from the back end. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name must be a valid and unique response header name and JSON-expression a valid JSON expression without the $ prefix.

$sel:selectionPattern:PutIntegrationResponse', putIntegrationResponse_selectionPattern - Specifies the selection pattern of a put integration response.

$sel:restApiId:PutIntegrationResponse', putIntegrationResponse_restApiId - [Required] The string identifier of the associated RestApi.

$sel:resourceId:PutIntegrationResponse', putIntegrationResponse_resourceId - [Required] Specifies a put integration response request's resource identifier.

$sel:httpMethod:PutIntegrationResponse', putIntegrationResponse_httpMethod - [Required] Specifies a put integration response request's HTTP method.

$sel:statusCode:PutIntegrationResponse', putIntegrationResponse_statusCode - [Required] Specifies the status code that is used to map the integration response to an existing MethodResponse.

Request Lenses

putIntegrationResponse_contentHandling :: Lens' PutIntegrationResponse (Maybe ContentHandlingStrategy)

Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the method response without modification.

putIntegrationResponse_responseTemplates :: Lens' PutIntegrationResponse (Maybe (HashMap Text Text))

Specifies a put integration response's templates.

putIntegrationResponse_responseParameters :: Lens' PutIntegrationResponse (Maybe (HashMap Text Text))

A key-value map specifying response parameters that are passed to the method response from the back end. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name must be a valid and unique response header name and JSON-expression a valid JSON expression without the $ prefix.

putIntegrationResponse_selectionPattern :: Lens' PutIntegrationResponse (Maybe Text)

Specifies the selection pattern of a put integration response.

putIntegrationResponse_restApiId :: Lens' PutIntegrationResponse Text

Required

The string identifier of the associated RestApi.

putIntegrationResponse_resourceId :: Lens' PutIntegrationResponse Text

Required

Specifies a put integration response request's resource identifier.

putIntegrationResponse_httpMethod :: Lens' PutIntegrationResponse Text

Required

Specifies a put integration response request's HTTP method.

putIntegrationResponse_statusCode :: Lens' PutIntegrationResponse Text

Required

Specifies the status code that is used to map the integration response to an existing MethodResponse.

Destructuring the Response

data IntegrationResponse

Represents an integration response. The status code must map to an existing MethodResponse, and parameters and templates can be used to transform the back-end response.

Creating an API

See: newIntegrationResponse smart constructor.

Constructors

IntegrationResponse'

Fields

contentHandling :: Maybe ContentHandlingStrategy

Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the method response without modification.

responseTemplates :: Maybe (HashMap Text Text)

Specifies the templates used to transform the integration response body. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

responseParameters :: Maybe (HashMap Text Text)

A key-value map specifying response parameters that are passed to the method response from the back end. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

statusCode :: Maybe Text

Specifies the status code that is used to map the integration response to an existing MethodResponse.

selectionPattern :: Maybe Text

Specifies the regular expression (regex) pattern used to choose an integration response based on the response from the back end. For example, if the success response returns nothing and the error response returns some string, you could use the .+ regex to match error response. However, make sure that the error response does not contain any newline (\n) character in such cases. If the back end is an AWS Lambda function, the AWS Lambda function error header is matched. For all other HTTP and AWS back ends, the HTTP status code is matched.

newIntegrationResponse :: IntegrationResponse

Create a value of IntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentHandling:IntegrationResponse', integrationResponse_contentHandling - Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the method response without modification.

$sel:responseTemplates:IntegrationResponse', integrationResponse_responseTemplates - Specifies the templates used to transform the integration response body. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

$sel:responseParameters:IntegrationResponse', integrationResponse_responseParameters - A key-value map specifying response parameters that are passed to the method response from the back end. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

$sel:statusCode:IntegrationResponse', integrationResponse_statusCode - Specifies the status code that is used to map the integration response to an existing MethodResponse.

$sel:selectionPattern:IntegrationResponse', integrationResponse_selectionPattern - Specifies the regular expression (regex) pattern used to choose an integration response based on the response from the back end. For example, if the success response returns nothing and the error response returns some string, you could use the .+ regex to match error response. However, make sure that the error response does not contain any newline (\n) character in such cases. If the back end is an AWS Lambda function, the AWS Lambda function error header is matched. For all other HTTP and AWS back ends, the HTTP status code is matched.

Response Lenses

integrationResponse_contentHandling :: Lens' IntegrationResponse (Maybe ContentHandlingStrategy)

Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the method response without modification.

integrationResponse_responseTemplates :: Lens' IntegrationResponse (Maybe (HashMap Text Text))

Specifies the templates used to transform the integration response body. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

integrationResponse_responseParameters :: Lens' IntegrationResponse (Maybe (HashMap Text Text))

A key-value map specifying response parameters that are passed to the method response from the back end. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

integrationResponse_statusCode :: Lens' IntegrationResponse (Maybe Text)

Specifies the status code that is used to map the integration response to an existing MethodResponse.

integrationResponse_selectionPattern :: Lens' IntegrationResponse (Maybe Text)

Specifies the regular expression (regex) pattern used to choose an integration response based on the response from the back end. For example, if the success response returns nothing and the error response returns some string, you could use the .+ regex to match error response. However, make sure that the error response does not contain any newline (\n) character in such cases. If the back end is an AWS Lambda function, the AWS Lambda function error header is matched. For all other HTTP and AWS back ends, the HTTP status code is matched.