Network.AWS.APIGateway.PutRestApi

Description

A feature of the API Gateway control service for updating an existing API with an input of external API definitions. The update can take the form of merging the supplied definition into the existing API or overwriting the existing API.

Creating a Request

data PutRestApi

A PUT request to update an existing API, with external API definitions specified as the request body.

See: newPutRestApi smart constructor.

Constructors

PutRestApi'

Fields

mode :: Maybe PutMode

The mode query parameter to specify the update mode. Valid values are "merge" and "overwrite". By default, the update mode is "merge".

failOnWarnings :: Maybe Bool

A query parameter to indicate whether to rollback the API update (true) or not (false) when a warning is encountered. The default value is false.

parameters :: Maybe (HashMap Text Text)

Custom header parameters as part of the request. For example, to exclude DocumentationParts from an imported API, set ignore=documentation as a parameters value, as in the AWS CLI command of aws apigateway import-rest-api --parameters ignore=documentation --body 'file:///path/to/imported-api-body.json'.

restApiId :: Text

Required

The string identifier of the associated RestApi.

body :: ByteString

Required

The PUT request body containing external API definitions. Currently, only OpenAPI definition JSON/YAML files are supported. The maximum size of the API definition file is 6MB.

newPutRestApi

Arguments

:: Text
$sel:restApiId:PutRestApi'

-> ByteString
$sel:body:PutRestApi'

-> PutRestApi

Create a value of PutRestApi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mode:PutRestApi', putRestApi_mode - The mode query parameter to specify the update mode. Valid values are "merge" and "overwrite". By default, the update mode is "merge".

$sel:failOnWarnings:PutRestApi', putRestApi_failOnWarnings - A query parameter to indicate whether to rollback the API update (true) or not (false) when a warning is encountered. The default value is false.

$sel:parameters:PutRestApi', putRestApi_parameters - Custom header parameters as part of the request. For example, to exclude DocumentationParts from an imported API, set ignore=documentation as a parameters value, as in the AWS CLI command of aws apigateway import-rest-api --parameters ignore=documentation --body 'file:///path/to/imported-api-body.json'.

$sel:restApiId:PutRestApi', putRestApi_restApiId - [Required] The string identifier of the associated RestApi.

$sel:body:PutRestApi', putRestApi_body - [Required] The PUT request body containing external API definitions. Currently, only OpenAPI definition JSON/YAML files are supported. The maximum size of the API definition file is 6MB.

Request Lenses

putRestApi_mode :: Lens' PutRestApi (Maybe PutMode)

The mode query parameter to specify the update mode. Valid values are "merge" and "overwrite". By default, the update mode is "merge".

putRestApi_failOnWarnings :: Lens' PutRestApi (Maybe Bool)

A query parameter to indicate whether to rollback the API update (true) or not (false) when a warning is encountered. The default value is false.

putRestApi_parameters :: Lens' PutRestApi (Maybe (HashMap Text Text))

Custom header parameters as part of the request. For example, to exclude DocumentationParts from an imported API, set ignore=documentation as a parameters value, as in the AWS CLI command of aws apigateway import-rest-api --parameters ignore=documentation --body 'file:///path/to/imported-api-body.json'.

putRestApi_restApiId :: Lens' PutRestApi Text

Required

The string identifier of the associated RestApi.

putRestApi_body :: Lens' PutRestApi ByteString

Required

The PUT request body containing external API definitions. Currently, only OpenAPI definition JSON/YAML files are supported. The maximum size of the API definition file is 6MB.

Destructuring the Response

data RestApi

Represents a REST API.

Create an API

See: newRestApi smart constructor.

Constructors

RestApi'

Fields

createdDate :: Maybe POSIX

The timestamp when the API was created.

warnings :: Maybe [Text]

The warning messages reported when failonwarnings is turned on during API import.

endpointConfiguration :: Maybe EndpointConfiguration

The endpoint configuration of this RestApi showing the endpoint types of the API.

binaryMediaTypes :: Maybe [Text]

The list of binary media types supported by the RestApi. By default, the RestApi supports only UTF-8-encoded text payloads.

id :: Maybe Text

The API's identifier. This identifier is unique across all of your APIs in API Gateway.

version :: Maybe Text

A version identifier for the API.

name :: Maybe Text

The API's name.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

description :: Maybe Text

The API's description.

policy :: Maybe Text

A stringified JSON policy document that applies to this RestApi regardless of the caller and Method configuration.

disableExecuteApiEndpoint :: Maybe Bool

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

minimumCompressionSize :: Maybe Int

A nullable integer that is used to enable compression (with non-negative between 0 and 10485760 (10M) bytes, inclusive) or disable compression (with a null value) on an API. When compression is enabled, compression or decompression is not applied on the payload if the payload size is smaller than this value. Setting it to zero allows compression for any payload size.

apiKeySource :: Maybe ApiKeySourceType

The source of the API key for metering requests according to a usage plan. Valid values are:

HEADER to read the API key from the X-API-Key header of a request.

AUTHORIZER to read the API key from the UsageIdentifierKey from a custom authorizer.

newRestApi :: RestApi

Create a value of RestApi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:RestApi', restApi_createdDate - The timestamp when the API was created.

$sel:warnings:RestApi', restApi_warnings - The warning messages reported when failonwarnings is turned on during API import.

$sel:endpointConfiguration:RestApi', restApi_endpointConfiguration - The endpoint configuration of this RestApi showing the endpoint types of the API.

$sel:binaryMediaTypes:RestApi', restApi_binaryMediaTypes - The list of binary media types supported by the RestApi. By default, the RestApi supports only UTF-8-encoded text payloads.

$sel:id:RestApi', restApi_id - The API's identifier. This identifier is unique across all of your APIs in API Gateway.

$sel:version:RestApi', restApi_version - A version identifier for the API.

$sel:name:RestApi', restApi_name - The API's name.

$sel:tags:RestApi', restApi_tags - The collection of tags. Each tag element is associated with a given resource.

$sel:description:RestApi', restApi_description - The API's description.

$sel:policy:RestApi', restApi_policy - A stringified JSON policy document that applies to this RestApi regardless of the caller and Method configuration.

$sel:disableExecuteApiEndpoint:RestApi', restApi_disableExecuteApiEndpoint - Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

$sel:minimumCompressionSize:RestApi', restApi_minimumCompressionSize - A nullable integer that is used to enable compression (with non-negative between 0 and 10485760 (10M) bytes, inclusive) or disable compression (with a null value) on an API. When compression is enabled, compression or decompression is not applied on the payload if the payload size is smaller than this value. Setting it to zero allows compression for any payload size.

$sel:apiKeySource:RestApi', restApi_apiKeySource - The source of the API key for metering requests according to a usage plan. Valid values are:

HEADER to read the API key from the X-API-Key header of a request.

AUTHORIZER to read the API key from the UsageIdentifierKey from a custom authorizer.

Response Lenses

restApi_createdDate :: Lens' RestApi (Maybe UTCTime)

The timestamp when the API was created.

restApi_warnings :: Lens' RestApi (Maybe [Text])

The warning messages reported when failonwarnings is turned on during API import.

restApi_endpointConfiguration :: Lens' RestApi (Maybe EndpointConfiguration)

The endpoint configuration of this RestApi showing the endpoint types of the API.

restApi_binaryMediaTypes :: Lens' RestApi (Maybe [Text])

The list of binary media types supported by the RestApi. By default, the RestApi supports only UTF-8-encoded text payloads.

restApi_id :: Lens' RestApi (Maybe Text)

The API's identifier. This identifier is unique across all of your APIs in API Gateway.

restApi_version :: Lens' RestApi (Maybe Text)

A version identifier for the API.

restApi_name :: Lens' RestApi (Maybe Text)

The API's name.

restApi_tags :: Lens' RestApi (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.

restApi_description :: Lens' RestApi (Maybe Text)

The API's description.

restApi_policy :: Lens' RestApi (Maybe Text)

A stringified JSON policy document that applies to this RestApi regardless of the caller and Method configuration.

restApi_disableExecuteApiEndpoint :: Lens' RestApi (Maybe Bool)

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

restApi_minimumCompressionSize :: Lens' RestApi (Maybe Int)

A nullable integer that is used to enable compression (with non-negative between 0 and 10485760 (10M) bytes, inclusive) or disable compression (with a null value) on an API. When compression is enabled, compression or decompression is not applied on the payload if the payload size is smaller than this value. Setting it to zero allows compression for any payload size.

restApi_apiKeySource :: Lens' RestApi (Maybe ApiKeySourceType)

The source of the API key for metering requests according to a usage plan. Valid values are:

HEADER to read the API key from the X-API-Key header of a request.

AUTHORIZER to read the API key from the UsageIdentifierKey from a custom authorizer.