Network.AWS.APIGateway.TestInvokeAuthorizer

Description

Simulate the execution of an Authorizer in your RestApi with headers, parameters, and an incoming request body.

Use Lambda Function as Authorizer Use Cognito User Pool as Authorizer

Creating a Request

data TestInvokeAuthorizer

Make a request to simulate the execution of an Authorizer.

See: newTestInvokeAuthorizer smart constructor.

Constructors

TestInvokeAuthorizer'

Fields

headers :: Maybe (HashMap Text Text)

Required

A key-value map of headers to simulate an incoming invocation request. This is where the incoming authorization token, or identity source, should be specified.

stageVariables :: Maybe (HashMap Text Text)

A key-value map of stage variables to simulate an invocation on a deployed Stage.

additionalContext :: Maybe (HashMap Text Text)

Optional

A key-value map of additional context variables.

body :: Maybe Text

Optional

The simulated request body of an incoming invocation request.

multiValueHeaders :: Maybe (HashMap Text [Text])

Optional

The headers as a map from string to list of values to simulate an incoming invocation request. This is where the incoming authorization token, or identity source, may be specified.

pathWithQueryString :: Maybe Text

Optional

The URI path, including query string, of the simulated invocation request. Use this to specify path parameters and query string parameters.

restApiId :: Text

Required

The string identifier of the associated RestApi.

authorizerId :: Text

Required

Specifies a test invoke authorizer request's Authorizer ID.

newTestInvokeAuthorizer

Arguments

:: Text
$sel:restApiId:TestInvokeAuthorizer'

-> Text
$sel:authorizerId:TestInvokeAuthorizer'

-> TestInvokeAuthorizer

Create a value of TestInvokeAuthorizer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:headers:TestInvokeAuthorizer', testInvokeAuthorizer_headers - [Required] A key-value map of headers to simulate an incoming invocation request. This is where the incoming authorization token, or identity source, should be specified.

$sel:stageVariables:TestInvokeAuthorizer', testInvokeAuthorizer_stageVariables - A key-value map of stage variables to simulate an invocation on a deployed Stage.

$sel:additionalContext:TestInvokeAuthorizer', testInvokeAuthorizer_additionalContext - [Optional] A key-value map of additional context variables.

$sel:body:TestInvokeAuthorizer', testInvokeAuthorizer_body - [Optional] The simulated request body of an incoming invocation request.

$sel:multiValueHeaders:TestInvokeAuthorizer', testInvokeAuthorizer_multiValueHeaders - [Optional] The headers as a map from string to list of values to simulate an incoming invocation request. This is where the incoming authorization token, or identity source, may be specified.

$sel:pathWithQueryString:TestInvokeAuthorizer', testInvokeAuthorizer_pathWithQueryString - [Optional] The URI path, including query string, of the simulated invocation request. Use this to specify path parameters and query string parameters.

$sel:restApiId:TestInvokeAuthorizer', testInvokeAuthorizer_restApiId - [Required] The string identifier of the associated RestApi.

$sel:authorizerId:TestInvokeAuthorizer', testInvokeAuthorizer_authorizerId - [Required] Specifies a test invoke authorizer request's Authorizer ID.

Request Lenses

testInvokeAuthorizer_headers :: Lens' TestInvokeAuthorizer (Maybe (HashMap Text Text))

Required

A key-value map of headers to simulate an incoming invocation request. This is where the incoming authorization token, or identity source, should be specified.

testInvokeAuthorizer_stageVariables :: Lens' TestInvokeAuthorizer (Maybe (HashMap Text Text))

A key-value map of stage variables to simulate an invocation on a deployed Stage.

testInvokeAuthorizer_additionalContext :: Lens' TestInvokeAuthorizer (Maybe (HashMap Text Text))

Optional

A key-value map of additional context variables.

testInvokeAuthorizer_body :: Lens' TestInvokeAuthorizer (Maybe Text)

Optional

The simulated request body of an incoming invocation request.

testInvokeAuthorizer_multiValueHeaders :: Lens' TestInvokeAuthorizer (Maybe (HashMap Text [Text]))

Optional

The headers as a map from string to list of values to simulate an incoming invocation request. This is where the incoming authorization token, or identity source, may be specified.

testInvokeAuthorizer_pathWithQueryString :: Lens' TestInvokeAuthorizer (Maybe Text)

Optional

The URI path, including query string, of the simulated invocation request. Use this to specify path parameters and query string parameters.

testInvokeAuthorizer_restApiId :: Lens' TestInvokeAuthorizer Text

Required

The string identifier of the associated RestApi.

testInvokeAuthorizer_authorizerId :: Lens' TestInvokeAuthorizer Text

Required

Specifies a test invoke authorizer request's Authorizer ID.

Destructuring the Response

data TestInvokeAuthorizerResponse

Represents the response of the test invoke request for a custom Authorizer

See: newTestInvokeAuthorizerResponse smart constructor.

Constructors

TestInvokeAuthorizerResponse'

Fields

claims :: Maybe (HashMap Text Text)

The open identity claims, with any supported custom attributes, returned from the Cognito Your User Pool configured for the API.

clientStatus :: Maybe Int

The HTTP status code that the client would have received. Value is 0 if the authorizer succeeded.

principalId :: Maybe Text

The principal identity returned by the Authorizer

log :: Maybe Text

The API Gateway execution log for the test authorizer request.

authorization :: Maybe (HashMap Text [Text])

policy :: Maybe Text

The JSON policy document returned by the Authorizer

latency :: Maybe Integer

The execution latency of the test authorizer request.

httpStatus :: Int

The response's http status code.

newTestInvokeAuthorizerResponse

Arguments

:: Int
$sel:httpStatus:TestInvokeAuthorizerResponse'

-> TestInvokeAuthorizerResponse

Create a value of TestInvokeAuthorizerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:claims:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_claims - The open identity claims, with any supported custom attributes, returned from the Cognito Your User Pool configured for the API.

$sel:clientStatus:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_clientStatus - The HTTP status code that the client would have received. Value is 0 if the authorizer succeeded.

$sel:principalId:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_principalId - The principal identity returned by the Authorizer

$sel:log:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_log - The API Gateway execution log for the test authorizer request.

$sel:authorization:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_authorization - Undocumented member.

$sel:policy:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_policy - The JSON policy document returned by the Authorizer

$sel:latency:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_latency - The execution latency of the test authorizer request.

$sel:httpStatus:TestInvokeAuthorizerResponse', testInvokeAuthorizerResponse_httpStatus - The response's http status code.

Response Lenses

testInvokeAuthorizerResponse_claims :: Lens' TestInvokeAuthorizerResponse (Maybe (HashMap Text Text))

The open identity claims, with any supported custom attributes, returned from the Cognito Your User Pool configured for the API.

testInvokeAuthorizerResponse_clientStatus :: Lens' TestInvokeAuthorizerResponse (Maybe Int)

The HTTP status code that the client would have received. Value is 0 if the authorizer succeeded.

testInvokeAuthorizerResponse_principalId :: Lens' TestInvokeAuthorizerResponse (Maybe Text)

The principal identity returned by the Authorizer

testInvokeAuthorizerResponse_log :: Lens' TestInvokeAuthorizerResponse (Maybe Text)

The API Gateway execution log for the test authorizer request.

testInvokeAuthorizerResponse_authorization :: Lens' TestInvokeAuthorizerResponse (Maybe (HashMap Text [Text]))

Undocumented member.

testInvokeAuthorizerResponse_policy :: Lens' TestInvokeAuthorizerResponse (Maybe Text)

The JSON policy document returned by the Authorizer

testInvokeAuthorizerResponse_latency :: Lens' TestInvokeAuthorizerResponse (Maybe Integer)

The execution latency of the test authorizer request.

testInvokeAuthorizerResponse_httpStatus :: Lens' TestInvokeAuthorizerResponse Int

The response's http status code.