Network.AWS.APIGateway.Types.AccessLogSettings

Description

Documentation

data AccessLogSettings

Access log settings, including the access log format and access log destination ARN.

See: newAccessLogSettings smart constructor.

Constructors

AccessLogSettings'

Fields

destinationArn :: Maybe Text

The Amazon Resource Name (ARN) of the CloudWatch Logs log group or Kinesis Data Firehose delivery stream to receive access logs. If you specify a Kinesis Data Firehose delivery stream, the stream name must begin with amazon-apigateway-.

format :: Maybe Text

A single line format of the access logs of data, as specified by selected $context variables. The format must include at least $context.requestId.

newAccessLogSettings :: AccessLogSettings

Create a value of AccessLogSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationArn:AccessLogSettings', accessLogSettings_destinationArn - The Amazon Resource Name (ARN) of the CloudWatch Logs log group or Kinesis Data Firehose delivery stream to receive access logs. If you specify a Kinesis Data Firehose delivery stream, the stream name must begin with amazon-apigateway-.

$sel:format:AccessLogSettings', accessLogSettings_format - A single line format of the access logs of data, as specified by selected $context variables. The format must include at least $context.requestId.

accessLogSettings_destinationArn :: Lens' AccessLogSettings (Maybe Text)

The Amazon Resource Name (ARN) of the CloudWatch Logs log group or Kinesis Data Firehose delivery stream to receive access logs. If you specify a Kinesis Data Firehose delivery stream, the stream name must begin with amazon-apigateway-.

accessLogSettings_format :: Lens' AccessLogSettings (Maybe Text)

A single line format of the access logs of data, as specified by selected $context variables. The format must include at least $context.requestId.