Network.AWS.APIGateway.Types.CanarySettings

Description

Documentation

data CanarySettings

Configuration settings of a canary deployment.

See: newCanarySettings smart constructor.

Constructors

CanarySettings'

Fields

deploymentId :: Maybe Text

The ID of the canary deployment.

percentTraffic :: Maybe Double

The percent (0-100) of traffic diverted to a canary deployment.

useStageCache :: Maybe Bool

A Boolean flag to indicate whether the canary deployment uses the stage cache or not.

stageVariableOverrides :: Maybe (HashMap Text Text)

Stage variables overridden for a canary release deployment, including new stage variables introduced in the canary. These stage variables are represented as a string-to-string map between stage variable names and their values.

newCanarySettings :: CanarySettings

Create a value of CanarySettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:CanarySettings', canarySettings_deploymentId - The ID of the canary deployment.

$sel:percentTraffic:CanarySettings', canarySettings_percentTraffic - The percent (0-100) of traffic diverted to a canary deployment.

$sel:useStageCache:CanarySettings', canarySettings_useStageCache - A Boolean flag to indicate whether the canary deployment uses the stage cache or not.

$sel:stageVariableOverrides:CanarySettings', canarySettings_stageVariableOverrides - Stage variables overridden for a canary release deployment, including new stage variables introduced in the canary. These stage variables are represented as a string-to-string map between stage variable names and their values.

canarySettings_deploymentId :: Lens' CanarySettings (Maybe Text)

The ID of the canary deployment.

canarySettings_percentTraffic :: Lens' CanarySettings (Maybe Double)

The percent (0-100) of traffic diverted to a canary deployment.

canarySettings_useStageCache :: Lens' CanarySettings (Maybe Bool)

A Boolean flag to indicate whether the canary deployment uses the stage cache or not.

canarySettings_stageVariableOverrides :: Lens' CanarySettings (Maybe (HashMap Text Text))

Stage variables overridden for a canary release deployment, including new stage variables introduced in the canary. These stage variables are represented as a string-to-string map between stage variable names and their values.