Network.AWS.APIGateway.Types.MethodSnapshot

Description

Documentation

data MethodSnapshot

Represents a summary of a Method resource, given a particular date and time.

See: newMethodSnapshot smart constructor.

Constructors

MethodSnapshot'

Fields

apiKeyRequired :: Maybe Bool

Specifies whether the method requires a valid ApiKey.

authorizationType :: Maybe Text

The method's authorization type. Valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, CUSTOM for using a custom authorizer, or COGNITO_USER_POOLS for using a Cognito user pool.

newMethodSnapshot :: MethodSnapshot

Create a value of MethodSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiKeyRequired:MethodSnapshot', methodSnapshot_apiKeyRequired - Specifies whether the method requires a valid ApiKey.

$sel:authorizationType:MethodSnapshot', methodSnapshot_authorizationType - The method's authorization type. Valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, CUSTOM for using a custom authorizer, or COGNITO_USER_POOLS for using a Cognito user pool.

methodSnapshot_apiKeyRequired :: Lens' MethodSnapshot (Maybe Bool)

Specifies whether the method requires a valid ApiKey.

methodSnapshot_authorizationType :: Lens' MethodSnapshot (Maybe Text)

The method's authorization type. Valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, CUSTOM for using a custom authorizer, or COGNITO_USER_POOLS for using a Cognito user pool.