Network.AWS.APIGateway.Types.Model

Description

Documentation

data Model

Represents the data structure of a method's request or response payload.

A request model defines the data structure of the client-supplied request payload. A response model defines the data structure of the response payload returned by the back end. Although not required, models are useful for mapping payloads between the front end and back end.

A model is used for generating an API's SDK, validating the input request body, and creating a skeletal mapping template.

Method, MethodResponse, Models and Mappings

See: newModel smart constructor.

Constructors

Model'

Fields

schema :: Maybe Text

The schema for the model. For application/json models, this should be JSON schema draft 4 model. Do not include "\*/" characters in the description of any properties because such "\*/" characters may be interpreted as the closing marker for comments in some languages, such as Java or JavaScript, causing the installation of your API's SDK generated by API Gateway to fail.

contentType :: Maybe Text

The content-type for the model.

id :: Maybe Text

The identifier for the model resource.

name :: Maybe Text

The name of the model. Must be an alphanumeric string.

description :: Maybe Text

The description of the model.

newModel :: Model

Create a value of Model with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schema:Model', model_schema - The schema for the model. For application/json models, this should be JSON schema draft 4 model. Do not include "\*/" characters in the description of any properties because such "\*/" characters may be interpreted as the closing marker for comments in some languages, such as Java or JavaScript, causing the installation of your API's SDK generated by API Gateway to fail.

$sel:contentType:Model', model_contentType - The content-type for the model.

$sel:id:Model', model_id - The identifier for the model resource.

$sel:name:Model', model_name - The name of the model. Must be an alphanumeric string.

$sel:description:Model', model_description - The description of the model.

model_schema :: Lens' Model (Maybe Text)

The schema for the model. For application/json models, this should be JSON schema draft 4 model. Do not include "\*/" characters in the description of any properties because such "\*/" characters may be interpreted as the closing marker for comments in some languages, such as Java or JavaScript, causing the installation of your API's SDK generated by API Gateway to fail.

model_contentType :: Lens' Model (Maybe Text)

The content-type for the model.

model_id :: Lens' Model (Maybe Text)

The identifier for the model resource.

model_name :: Lens' Model (Maybe Text)

The name of the model. Must be an alphanumeric string.

model_description :: Lens' Model (Maybe Text)

The description of the model.