Network.AWS.APIGateway.Types.RestApi

Description

Documentation

data RestApi

Represents a REST API.

Create an API

See: newRestApi smart constructor.

Constructors

RestApi'

Fields

minimumCompressionSize :: Maybe Int

A nullable integer that is used to enable compression (with non-negative between 0 and 10485760 (10M) bytes, inclusive) or disable compression (with a null value) on an API. When compression is enabled, compression or decompression is not applied on the payload if the payload size is smaller than this value. Setting it to zero allows compression for any payload size.

disableExecuteApiEndpoint :: Maybe Bool

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

binaryMediaTypes :: Maybe [Text]

The list of binary media types supported by the RestApi. By default, the RestApi supports only UTF-8-encoded text payloads.

warnings :: Maybe [Text]

The warning messages reported when failonwarnings is turned on during API import.

createdDate :: Maybe POSIX

The timestamp when the API was created.

name :: Maybe Text

The API's name.

version :: Maybe Text

A version identifier for the API.

apiKeySource :: Maybe ApiKeySourceType

The source of the API key for metering requests according to a usage plan. Valid values are:

HEADER to read the API key from the X-API-Key header of a request.

AUTHORIZER to read the API key from the UsageIdentifierKey from a custom authorizer.

id :: Maybe Text

The API's identifier. This identifier is unique across all of your APIs in API Gateway.

policy :: Maybe Text

A stringified JSON policy document that applies to this RestApi regardless of the caller and Method configuration.

endpointConfiguration :: Maybe EndpointConfiguration

The endpoint configuration of this RestApi showing the endpoint types of the API.

description :: Maybe Text

The API's description.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

newRestApi :: RestApi

Create a value of RestApi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minimumCompressionSize:RestApi', restApi_minimumCompressionSize - A nullable integer that is used to enable compression (with non-negative between 0 and 10485760 (10M) bytes, inclusive) or disable compression (with a null value) on an API. When compression is enabled, compression or decompression is not applied on the payload if the payload size is smaller than this value. Setting it to zero allows compression for any payload size.

$sel:disableExecuteApiEndpoint:RestApi', restApi_disableExecuteApiEndpoint - Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

$sel:binaryMediaTypes:RestApi', restApi_binaryMediaTypes - The list of binary media types supported by the RestApi. By default, the RestApi supports only UTF-8-encoded text payloads.

$sel:warnings:RestApi', restApi_warnings - The warning messages reported when failonwarnings is turned on during API import.

$sel:createdDate:RestApi', restApi_createdDate - The timestamp when the API was created.

$sel:name:RestApi', restApi_name - The API's name.

$sel:version:RestApi', restApi_version - A version identifier for the API.

$sel:apiKeySource:RestApi', restApi_apiKeySource - The source of the API key for metering requests according to a usage plan. Valid values are:

HEADER to read the API key from the X-API-Key header of a request.

AUTHORIZER to read the API key from the UsageIdentifierKey from a custom authorizer.

$sel:id:RestApi', restApi_id - The API's identifier. This identifier is unique across all of your APIs in API Gateway.

$sel:policy:RestApi', restApi_policy - A stringified JSON policy document that applies to this RestApi regardless of the caller and Method configuration.

$sel:endpointConfiguration:RestApi', restApi_endpointConfiguration - The endpoint configuration of this RestApi showing the endpoint types of the API.

$sel:description:RestApi', restApi_description - The API's description.

$sel:tags:RestApi', restApi_tags - The collection of tags. Each tag element is associated with a given resource.

restApi_minimumCompressionSize :: Lens' RestApi (Maybe Int)

A nullable integer that is used to enable compression (with non-negative between 0 and 10485760 (10M) bytes, inclusive) or disable compression (with a null value) on an API. When compression is enabled, compression or decompression is not applied on the payload if the payload size is smaller than this value. Setting it to zero allows compression for any payload size.

restApi_disableExecuteApiEndpoint :: Lens' RestApi (Maybe Bool)

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

restApi_binaryMediaTypes :: Lens' RestApi (Maybe [Text])

The list of binary media types supported by the RestApi. By default, the RestApi supports only UTF-8-encoded text payloads.

restApi_warnings :: Lens' RestApi (Maybe [Text])

The warning messages reported when failonwarnings is turned on during API import.

restApi_createdDate :: Lens' RestApi (Maybe UTCTime)

The timestamp when the API was created.

restApi_name :: Lens' RestApi (Maybe Text)

The API's name.

restApi_version :: Lens' RestApi (Maybe Text)

A version identifier for the API.

restApi_apiKeySource :: Lens' RestApi (Maybe ApiKeySourceType)

The source of the API key for metering requests according to a usage plan. Valid values are:

HEADER to read the API key from the X-API-Key header of a request.

AUTHORIZER to read the API key from the UsageIdentifierKey from a custom authorizer.

restApi_id :: Lens' RestApi (Maybe Text)

The API's identifier. This identifier is unique across all of your APIs in API Gateway.

restApi_policy :: Lens' RestApi (Maybe Text)

A stringified JSON policy document that applies to this RestApi regardless of the caller and Method configuration.

restApi_endpointConfiguration :: Lens' RestApi (Maybe EndpointConfiguration)

The endpoint configuration of this RestApi showing the endpoint types of the API.

restApi_description :: Lens' RestApi (Maybe Text)

The API's description.

restApi_tags :: Lens' RestApi (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.