Network.AWS.APIGateway.Types.Usage

Description

Documentation

data Usage

Represents the usage data of a usage plan.

Create and Use Usage Plans, Manage Usage in a Usage Plan

See: newUsage smart constructor.

Constructors

Usage'

Fields

startDate :: Maybe Text

The starting date of the usage data.

items :: Maybe (HashMap Text [[Integer]])

The usage data, as daily logs of used and remaining quotas, over the specified time interval indexed over the API keys in a usage plan. For example, {..., "values" : { "{api_key}" : [ [0, 100], [10, 90], [100, 10]]}, where {api_key} stands for an API key value and the daily log entry is of the format [used quota, remaining quota].

position :: Maybe Text

endDate :: Maybe Text

The ending date of the usage data.

usagePlanId :: Maybe Text

The plan Id associated with this usage data.

newUsage :: Usage

Create a value of Usage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startDate:Usage', usage_startDate - The starting date of the usage data.

$sel:items:Usage', usage_items - The usage data, as daily logs of used and remaining quotas, over the specified time interval indexed over the API keys in a usage plan. For example, {..., "values" : { "{api_key}" : [ [0, 100], [10, 90], [100, 10]]}, where {api_key} stands for an API key value and the daily log entry is of the format [used quota, remaining quota].

$sel:position:Usage', usage_position - Undocumented member.

$sel:endDate:Usage', usage_endDate - The ending date of the usage data.

$sel:usagePlanId:Usage', usage_usagePlanId - The plan Id associated with this usage data.

usage_startDate :: Lens' Usage (Maybe Text)

The starting date of the usage data.

usage_items :: Lens' Usage (Maybe (HashMap Text [[Integer]]))

The usage data, as daily logs of used and remaining quotas, over the specified time interval indexed over the API keys in a usage plan. For example, {..., "values" : { "{api_key}" : [ [0, 100], [10, 90], [100, 10]]}, where {api_key} stands for an API key value and the daily log entry is of the format [used quota, remaining quota].

usage_position :: Lens' Usage (Maybe Text)

Undocumented member.

usage_endDate :: Lens' Usage (Maybe Text)

The ending date of the usage data.

usage_usagePlanId :: Lens' Usage (Maybe Text)

The plan Id associated with this usage data.