Network.AWS.APIGateway.UpdateApiKey

Description

Changes information about an ApiKey resource.

Creating a Request

data UpdateApiKey

A request to change information about an ApiKey resource.

See: newUpdateApiKey smart constructor.

Constructors

UpdateApiKey'

Fields

patchOperations :: Maybe [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

apiKey :: Text

Required

The identifier of the ApiKey resource to be updated.

newUpdateApiKey

Arguments

:: Text
$sel:apiKey:UpdateApiKey'

-> UpdateApiKey

Create a value of UpdateApiKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:patchOperations:UpdateApiKey', updateApiKey_patchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

$sel:apiKey:UpdateApiKey', updateApiKey_apiKey - [Required] The identifier of the ApiKey resource to be updated.

Request Lenses

updateApiKey_patchOperations :: Lens' UpdateApiKey (Maybe [PatchOperation])

A list of update operations to be applied to the specified resource and in the order specified in this list.

updateApiKey_apiKey :: Lens' UpdateApiKey Text

Required

The identifier of the ApiKey resource to be updated.

Destructuring the Response

data ApiKey

A resource that can be distributed to callers for executing Method resources that require an API key. API keys can be mapped to any Stage on any RestApi, which indicates that the callers with the API key can make requests to that stage.

Use API Keys

See: newApiKey smart constructor.

Constructors

ApiKey'

Fields

createdDate :: Maybe POSIX

The timestamp when the API Key was created.

customerId :: Maybe Text

An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

lastUpdatedDate :: Maybe POSIX

The timestamp when the API Key was last updated.

stageKeys :: Maybe [Text]

A list of Stage resources that are associated with the ApiKey resource.

id :: Maybe Text

The identifier of the API Key.

enabled :: Maybe Bool

Specifies whether the API Key can be used by callers.

name :: Maybe Text

The name of the API Key.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

description :: Maybe Text

The description of the API Key.

value :: Maybe Text

The value of the API Key.

newApiKey :: ApiKey

Create a value of ApiKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:ApiKey', apiKey_createdDate - The timestamp when the API Key was created.

$sel:customerId:ApiKey', apiKey_customerId - An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

$sel:lastUpdatedDate:ApiKey', apiKey_lastUpdatedDate - The timestamp when the API Key was last updated.

$sel:stageKeys:ApiKey', apiKey_stageKeys - A list of Stage resources that are associated with the ApiKey resource.

$sel:id:ApiKey', apiKey_id - The identifier of the API Key.

$sel:enabled:ApiKey', apiKey_enabled - Specifies whether the API Key can be used by callers.

$sel:name:ApiKey', apiKey_name - The name of the API Key.

$sel:tags:ApiKey', apiKey_tags - The collection of tags. Each tag element is associated with a given resource.

$sel:description:ApiKey', apiKey_description - The description of the API Key.

$sel:value:ApiKey', apiKey_value - The value of the API Key.

Response Lenses

apiKey_createdDate :: Lens' ApiKey (Maybe UTCTime)

The timestamp when the API Key was created.

apiKey_customerId :: Lens' ApiKey (Maybe Text)

An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.

apiKey_lastUpdatedDate :: Lens' ApiKey (Maybe UTCTime)

The timestamp when the API Key was last updated.

apiKey_stageKeys :: Lens' ApiKey (Maybe [Text])

A list of Stage resources that are associated with the ApiKey resource.

apiKey_id :: Lens' ApiKey (Maybe Text)

The identifier of the API Key.

apiKey_enabled :: Lens' ApiKey (Maybe Bool)

Specifies whether the API Key can be used by callers.

apiKey_name :: Lens' ApiKey (Maybe Text)

The name of the API Key.

apiKey_tags :: Lens' ApiKey (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.

apiKey_description :: Lens' ApiKey (Maybe Text)

The description of the API Key.

apiKey_value :: Lens' ApiKey (Maybe Text)

The value of the API Key.