Network.AWS.APIGateway.UpdateBasePathMapping

Description

Changes information about the BasePathMapping resource.

Creating a Request

data UpdateBasePathMapping

A request to change information about the BasePathMapping resource.

See: newUpdateBasePathMapping smart constructor.

Constructors

UpdateBasePathMapping'

Fields

patchOperations :: Maybe [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

domainName :: Text

Required

The domain name of the BasePathMapping resource to change.

basePath :: Text

Required

The base path of the BasePathMapping resource to change.

To specify an empty base path, set this parameter to '(none)'.

newUpdateBasePathMapping

Arguments

:: Text
$sel:domainName:UpdateBasePathMapping'

-> Text
$sel:basePath:UpdateBasePathMapping'

-> UpdateBasePathMapping

Create a value of UpdateBasePathMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:patchOperations:UpdateBasePathMapping', updateBasePathMapping_patchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

$sel:domainName:UpdateBasePathMapping', updateBasePathMapping_domainName - [Required] The domain name of the BasePathMapping resource to change.

$sel:basePath:UpdateBasePathMapping', updateBasePathMapping_basePath - [Required] The base path of the BasePathMapping resource to change.

To specify an empty base path, set this parameter to '(none)'.

Request Lenses

updateBasePathMapping_patchOperations :: Lens' UpdateBasePathMapping (Maybe [PatchOperation])

A list of update operations to be applied to the specified resource and in the order specified in this list.

updateBasePathMapping_domainName :: Lens' UpdateBasePathMapping Text

Required

The domain name of the BasePathMapping resource to change.

updateBasePathMapping_basePath :: Lens' UpdateBasePathMapping Text

Required

The base path of the BasePathMapping resource to change.

To specify an empty base path, set this parameter to '(none)'.

Destructuring the Response

data BasePathMapping

Represents the base path that callers of the API must provide as part of the URL after the domain name.

A custom domain name plus a BasePathMapping specification identifies a deployed RestApi in a given stage of the owner Account.

Use Custom Domain Names

See: newBasePathMapping smart constructor.

Constructors

BasePathMapping'

Fields

basePath :: Maybe Text

The base path name that callers of the API must provide as part of the URL after the domain name.

stage :: Maybe Text

The name of the associated stage.

restApiId :: Maybe Text

The string identifier of the associated RestApi.

newBasePathMapping :: BasePathMapping

Create a value of BasePathMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:basePath:BasePathMapping', basePathMapping_basePath - The base path name that callers of the API must provide as part of the URL after the domain name.

$sel:stage:BasePathMapping', basePathMapping_stage - The name of the associated stage.

$sel:restApiId:BasePathMapping', basePathMapping_restApiId - The string identifier of the associated RestApi.

Response Lenses

basePathMapping_basePath :: Lens' BasePathMapping (Maybe Text)

The base path name that callers of the API must provide as part of the URL after the domain name.

basePathMapping_stage :: Lens' BasePathMapping (Maybe Text)

The name of the associated stage.

basePathMapping_restApiId :: Lens' BasePathMapping (Maybe Text)

The string identifier of the associated RestApi.