Network.AWS.APIGateway.UpdateClientCertificate

Description

Changes information about an ClientCertificate resource.

Creating a Request

data UpdateClientCertificate

A request to change information about an ClientCertificate resource.

See: newUpdateClientCertificate smart constructor.

Constructors

UpdateClientCertificate'

Fields

patchOperations :: Maybe [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

clientCertificateId :: Text

Required

The identifier of the ClientCertificate resource to be updated.

newUpdateClientCertificate

Arguments

:: Text
$sel:clientCertificateId:UpdateClientCertificate'

-> UpdateClientCertificate

Create a value of UpdateClientCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:patchOperations:UpdateClientCertificate', updateClientCertificate_patchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

$sel:clientCertificateId:UpdateClientCertificate', updateClientCertificate_clientCertificateId - [Required] The identifier of the ClientCertificate resource to be updated.

Request Lenses

updateClientCertificate_patchOperations :: Lens' UpdateClientCertificate (Maybe [PatchOperation])

A list of update operations to be applied to the specified resource and in the order specified in this list.

updateClientCertificate_clientCertificateId :: Lens' UpdateClientCertificate Text

Required

The identifier of the ClientCertificate resource to be updated.

Destructuring the Response

data ClientCertificate

Represents a client certificate used to configure client-side SSL authentication while sending requests to the integration endpoint.

Client certificates are used to authenticate an API by the backend server. To authenticate an API client (or user), use IAM roles and policies, a custom Authorizer or an Amazon Cognito user pool.

Use Client-Side Certificate

See: newClientCertificate smart constructor.

Constructors

ClientCertificate'

Fields

pemEncodedCertificate :: Maybe Text

The PEM-encoded public key of the client certificate, which can be used to configure certificate authentication in the integration endpoint .

clientCertificateId :: Maybe Text

The identifier of the client certificate.

createdDate :: Maybe POSIX

The timestamp when the client certificate was created.

expirationDate :: Maybe POSIX

The timestamp when the client certificate will expire.

description :: Maybe Text

The description of the client certificate.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

newClientCertificate :: ClientCertificate

Create a value of ClientCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pemEncodedCertificate:ClientCertificate', clientCertificate_pemEncodedCertificate - The PEM-encoded public key of the client certificate, which can be used to configure certificate authentication in the integration endpoint .

$sel:clientCertificateId:ClientCertificate', clientCertificate_clientCertificateId - The identifier of the client certificate.

$sel:createdDate:ClientCertificate', clientCertificate_createdDate - The timestamp when the client certificate was created.

$sel:expirationDate:ClientCertificate', clientCertificate_expirationDate - The timestamp when the client certificate will expire.

$sel:description:ClientCertificate', clientCertificate_description - The description of the client certificate.

$sel:tags:ClientCertificate', clientCertificate_tags - The collection of tags. Each tag element is associated with a given resource.

Response Lenses

clientCertificate_pemEncodedCertificate :: Lens' ClientCertificate (Maybe Text)

The PEM-encoded public key of the client certificate, which can be used to configure certificate authentication in the integration endpoint .

clientCertificate_clientCertificateId :: Lens' ClientCertificate (Maybe Text)

The identifier of the client certificate.

clientCertificate_createdDate :: Lens' ClientCertificate (Maybe UTCTime)

The timestamp when the client certificate was created.

clientCertificate_expirationDate :: Lens' ClientCertificate (Maybe UTCTime)

The timestamp when the client certificate will expire.

clientCertificate_description :: Lens' ClientCertificate (Maybe Text)

The description of the client certificate.

clientCertificate_tags :: Lens' ClientCertificate (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.