Network.AWS.APIGateway.UpdateDeployment

Description

Changes information about a Deployment resource.

Creating a Request

data UpdateDeployment

Requests API Gateway to change information about a Deployment resource.

See: newUpdateDeployment smart constructor.

Constructors

UpdateDeployment'

Fields

patchOperations :: Maybe [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

restApiId :: Text

Required

The string identifier of the associated RestApi.

deploymentId :: Text

The replacement identifier for the Deployment resource to change information about.

newUpdateDeployment

Arguments

:: Text
$sel:restApiId:UpdateDeployment'

-> Text
$sel:deploymentId:UpdateDeployment'

-> UpdateDeployment

Create a value of UpdateDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:patchOperations:UpdateDeployment', updateDeployment_patchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

$sel:restApiId:UpdateDeployment', updateDeployment_restApiId - [Required] The string identifier of the associated RestApi.

$sel:deploymentId:UpdateDeployment', updateDeployment_deploymentId - The replacement identifier for the Deployment resource to change information about.

Request Lenses

updateDeployment_patchOperations :: Lens' UpdateDeployment (Maybe [PatchOperation])

A list of update operations to be applied to the specified resource and in the order specified in this list.

updateDeployment_restApiId :: Lens' UpdateDeployment Text

Required

The string identifier of the associated RestApi.

updateDeployment_deploymentId :: Lens' UpdateDeployment Text

The replacement identifier for the Deployment resource to change information about.

Destructuring the Response

data Deployment

An immutable representation of a RestApi resource that can be called by users using Stages. A deployment must be associated with a Stage for it to be callable over the Internet.

To create a deployment, call POST on the Deployments resource of a RestApi. To view, update, or delete a deployment, call GET, PATCH, or DELETE on the specified deployment resource (/restapis/{restapi_id}/deployments/{deployment_id}).

RestApi, Deployments, Stage, AWS CLI, AWS SDKs

See: newDeployment smart constructor.

Constructors

Deployment'

Fields

apiSummary :: Maybe (HashMap Text (HashMap Text MethodSnapshot))

A summary of the RestApi at the date and time that the deployment resource was created.

createdDate :: Maybe POSIX

The date and time that the deployment resource was created.

id :: Maybe Text

The identifier for the deployment resource.

description :: Maybe Text

The description for the deployment resource.

newDeployment :: Deployment

Create a value of Deployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiSummary:Deployment', deployment_apiSummary - A summary of the RestApi at the date and time that the deployment resource was created.

$sel:createdDate:Deployment', deployment_createdDate - The date and time that the deployment resource was created.

$sel:id:Deployment', deployment_id - The identifier for the deployment resource.

$sel:description:Deployment', deployment_description - The description for the deployment resource.

Response Lenses

deployment_apiSummary :: Lens' Deployment (Maybe (HashMap Text (HashMap Text MethodSnapshot)))

A summary of the RestApi at the date and time that the deployment resource was created.

deployment_createdDate :: Lens' Deployment (Maybe UTCTime)

The date and time that the deployment resource was created.

deployment_id :: Lens' Deployment (Maybe Text)

The identifier for the deployment resource.

deployment_description :: Lens' Deployment (Maybe Text)

The description for the deployment resource.