Network.AWS.APIGateway.UpdateGatewayResponse

Description

Updates a GatewayResponse of a specified response type on the given RestApi.

Creating a Request

data UpdateGatewayResponse

Updates a GatewayResponse of a specified response type on the given RestApi.

See: newUpdateGatewayResponse smart constructor.

Constructors

UpdateGatewayResponse'

Fields

patchOperations :: Maybe [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

restApiId :: Text

Required

The string identifier of the associated RestApi.

responseType :: GatewayResponseType

Required

The response type of the associated GatewayResponse.

newUpdateGatewayResponse

Arguments

:: Text
$sel:restApiId:UpdateGatewayResponse'

-> GatewayResponseType
$sel:responseType:UpdateGatewayResponse'

-> UpdateGatewayResponse

Create a value of UpdateGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:patchOperations:UpdateGatewayResponse', updateGatewayResponse_patchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

$sel:restApiId:UpdateGatewayResponse', updateGatewayResponse_restApiId - [Required] The string identifier of the associated RestApi.

$sel:responseType:UpdateGatewayResponse', updateGatewayResponse_responseType - [Required]

The response type of the associated GatewayResponse.

Request Lenses

updateGatewayResponse_patchOperations :: Lens' UpdateGatewayResponse (Maybe [PatchOperation])

A list of update operations to be applied to the specified resource and in the order specified in this list.

updateGatewayResponse_restApiId :: Lens' UpdateGatewayResponse Text

Required

The string identifier of the associated RestApi.

updateGatewayResponse_responseType :: Lens' UpdateGatewayResponse GatewayResponseType

Required

The response type of the associated GatewayResponse.

Destructuring the Response

data GatewayResponse

A gateway response of a given response type and status code, with optional response parameters and mapping templates.

For more information about valid gateway response types, see Gateway Response Types Supported by API Gateway

Example: Get a Gateway Response of a given response type

Request

This example shows how to get a gateway response of the MISSING_AUTHENTICATION_TOKEN type.

GET /restapis/o81lxisefl/gatewayresponses/MISSING_AUTHENTICATION_TOKEN HTTP/1.1 Host: beta-apigateway.us-east-1.amazonaws.com Content-Type: application/json X-Amz-Date: 20170503T202516Z Authorization: AWS4-HMAC-SHA256 Credential={access-key-id}/20170503/us-east-1/apigateway/aws4_request, SignedHeaders=content-type;host;x-amz-date, Signature=1b52460e3159c1a26cff29093855d50ea141c1c5b937528fecaf60f51129697a Cache-Control: no-cache Postman-Token: 3b2a1ce9-c848-2e26-2e2f-9c2caefbed45

The response type is specified as a URL path.

Response

The successful operation returns the 200 OK status code and a payload similar to the following:

{ "_links": { "curies": { "href": "http://docs.aws.amazon.com/apigateway/latest/developerguide/restapi-gatewayresponse-{rel}.html", "name": "gatewayresponse", "templated": true }, "self": { "href": "/restapis/o81lxisefl/gatewayresponses/MISSING_AUTHENTICATION_TOKEN" }, "gatewayresponse:delete": { "href": "/restapis/o81lxisefl/gatewayresponses/MISSING_AUTHENTICATION_TOKEN" }, "gatewayresponse:put": { "href": "/restapis/o81lxisefl/gatewayresponses/{response_type}", "templated": true }, "gatewayresponse:update": { "href": "/restapis/o81lxisefl/gatewayresponses/MISSING_AUTHENTICATION_TOKEN" } }, "defaultResponse": false, "responseParameters": { "gatewayresponse.header.x-request-path": "method.request.path.petId", "gatewayresponse.header.Access-Control-Allow-Origin": "'a.b.c'", "gatewayresponse.header.x-request-query": "method.request.querystring.q", "gatewayresponse.header.x-request-header": "method.request.header.Accept" }, "responseTemplates": { "application/json": "{\n \"message\": $context.error.messageString,\n \"type\": \"$context.error.responseType\",\n \"stage\": \"$context.stage\",\n \"resourcePath\": \"$context.resourcePath\",\n \"stageVariables.a\": \"$stageVariables.a\",\n \"statusCode\": \"'404'\"\n}" }, "responseType": "MISSING_AUTHENTICATION_TOKEN", "statusCode": "404" }

Customize Gateway Responses

See: newGatewayResponse smart constructor.

Constructors

GatewayResponse'

Fields

responseTemplates :: Maybe (HashMap Text Text)

Response templates of the GatewayResponse as a string-to-string map of key-value pairs.

responseParameters :: Maybe (HashMap Text Text)

Response parameters (paths, query strings and headers) of the GatewayResponse as a string-to-string map of key-value pairs.

statusCode :: Maybe Text

The HTTP status code for this GatewayResponse.

responseType :: Maybe GatewayResponseType

The response type of the associated GatewayResponse.

defaultResponse :: Maybe Bool

A Boolean flag to indicate whether this GatewayResponse is the default gateway response (true) or not (false). A default gateway response is one generated by API Gateway without any customization by an API developer.

newGatewayResponse :: GatewayResponse

Create a value of GatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:responseTemplates:GatewayResponse', gatewayResponse_responseTemplates - Response templates of the GatewayResponse as a string-to-string map of key-value pairs.

$sel:responseParameters:GatewayResponse', gatewayResponse_responseParameters - Response parameters (paths, query strings and headers) of the GatewayResponse as a string-to-string map of key-value pairs.

$sel:statusCode:GatewayResponse', gatewayResponse_statusCode - The HTTP status code for this GatewayResponse.

$sel:responseType:GatewayResponse', gatewayResponse_responseType - The response type of the associated GatewayResponse.

$sel:defaultResponse:GatewayResponse', gatewayResponse_defaultResponse - A Boolean flag to indicate whether this GatewayResponse is the default gateway response (true) or not (false). A default gateway response is one generated by API Gateway without any customization by an API developer.

Response Lenses

gatewayResponse_responseTemplates :: Lens' GatewayResponse (Maybe (HashMap Text Text))

Response templates of the GatewayResponse as a string-to-string map of key-value pairs.

gatewayResponse_responseParameters :: Lens' GatewayResponse (Maybe (HashMap Text Text))

Response parameters (paths, query strings and headers) of the GatewayResponse as a string-to-string map of key-value pairs.

gatewayResponse_statusCode :: Lens' GatewayResponse (Maybe Text)

The HTTP status code for this GatewayResponse.

gatewayResponse_responseType :: Lens' GatewayResponse (Maybe GatewayResponseType)

The response type of the associated GatewayResponse.

gatewayResponse_defaultResponse :: Lens' GatewayResponse (Maybe Bool)

A Boolean flag to indicate whether this GatewayResponse is the default gateway response (true) or not (false). A default gateway response is one generated by API Gateway without any customization by an API developer.