Network.AWS.APIGateway.UpdateModel

Description

Changes information about a model.

Creating a Request

data UpdateModel

Request to update an existing model in an existing RestApi resource.

See: newUpdateModel smart constructor.

Constructors

UpdateModel'

Fields

patchOperations :: Maybe [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

restApiId :: Text

Required

The string identifier of the associated RestApi.

modelName :: Text

Required

The name of the model to update.

newUpdateModel

Arguments

:: Text
$sel:restApiId:UpdateModel'

-> Text
$sel:modelName:UpdateModel'

-> UpdateModel

Create a value of UpdateModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:patchOperations:UpdateModel', updateModel_patchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

$sel:restApiId:UpdateModel', updateModel_restApiId - [Required] The string identifier of the associated RestApi.

$sel:modelName:UpdateModel', updateModel_modelName - [Required] The name of the model to update.

Request Lenses

updateModel_patchOperations :: Lens' UpdateModel (Maybe [PatchOperation])

A list of update operations to be applied to the specified resource and in the order specified in this list.

updateModel_restApiId :: Lens' UpdateModel Text

Required

The string identifier of the associated RestApi.

updateModel_modelName :: Lens' UpdateModel Text

Required

The name of the model to update.

Destructuring the Response

data Model

Represents the data structure of a method's request or response payload.

A request model defines the data structure of the client-supplied request payload. A response model defines the data structure of the response payload returned by the back end. Although not required, models are useful for mapping payloads between the front end and back end.

A model is used for generating an API's SDK, validating the input request body, and creating a skeletal mapping template.

Method, MethodResponse, Models and Mappings

See: newModel smart constructor.

Constructors

Model'

Fields

schema :: Maybe Text

The schema for the model. For application/json models, this should be JSON schema draft 4 model. Do not include "\*/" characters in the description of any properties because such "\*/" characters may be interpreted as the closing marker for comments in some languages, such as Java or JavaScript, causing the installation of your API's SDK generated by API Gateway to fail.

name :: Maybe Text

The name of the model. Must be an alphanumeric string.

id :: Maybe Text

The identifier for the model resource.

description :: Maybe Text

The description of the model.

contentType :: Maybe Text

The content-type for the model.

newModel :: Model

Create a value of Model with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schema:Model', model_schema - The schema for the model. For application/json models, this should be JSON schema draft 4 model. Do not include "\*/" characters in the description of any properties because such "\*/" characters may be interpreted as the closing marker for comments in some languages, such as Java or JavaScript, causing the installation of your API's SDK generated by API Gateway to fail.

$sel:name:Model', model_name - The name of the model. Must be an alphanumeric string.

$sel:id:Model', model_id - The identifier for the model resource.

$sel:description:Model', model_description - The description of the model.

$sel:contentType:Model', model_contentType - The content-type for the model.

Response Lenses

model_schema :: Lens' Model (Maybe Text)

The schema for the model. For application/json models, this should be JSON schema draft 4 model. Do not include "\*/" characters in the description of any properties because such "\*/" characters may be interpreted as the closing marker for comments in some languages, such as Java or JavaScript, causing the installation of your API's SDK generated by API Gateway to fail.

model_name :: Lens' Model (Maybe Text)

The name of the model. Must be an alphanumeric string.

model_id :: Lens' Model (Maybe Text)

The identifier for the model resource.

model_description :: Lens' Model (Maybe Text)

The description of the model.

model_contentType :: Lens' Model (Maybe Text)

The content-type for the model.