Network.AWS.APIGateway.UpdateRequestValidator

Description

Updates a RequestValidator of a given RestApi.

Creating a Request

data UpdateRequestValidator

Updates a RequestValidator of a given RestApi.

See: newUpdateRequestValidator smart constructor.

Constructors

UpdateRequestValidator'

Fields

patchOperations :: Maybe [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

restApiId :: Text

Required

The string identifier of the associated RestApi.

requestValidatorId :: Text

Required

The identifier of RequestValidator to be updated.

newUpdateRequestValidator

Arguments

:: Text
$sel:restApiId:UpdateRequestValidator'

-> Text
$sel:requestValidatorId:UpdateRequestValidator'

-> UpdateRequestValidator

Create a value of UpdateRequestValidator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:patchOperations:UpdateRequestValidator', updateRequestValidator_patchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

$sel:restApiId:UpdateRequestValidator', updateRequestValidator_restApiId - [Required] The string identifier of the associated RestApi.

$sel:requestValidatorId:UpdateRequestValidator', updateRequestValidator_requestValidatorId - [Required] The identifier of RequestValidator to be updated.

Request Lenses

updateRequestValidator_patchOperations :: Lens' UpdateRequestValidator (Maybe [PatchOperation])

A list of update operations to be applied to the specified resource and in the order specified in this list.

updateRequestValidator_restApiId :: Lens' UpdateRequestValidator Text

Required

The string identifier of the associated RestApi.

updateRequestValidator_requestValidatorId :: Lens' UpdateRequestValidator Text

Required

The identifier of RequestValidator to be updated.

Destructuring the Response

data RequestValidator

A set of validation rules for incoming Method requests.

In OpenAPI, a RequestValidator of an API is defined by the x-amazon-apigateway-request-validators.requestValidator object. It the referenced using the x-amazon-apigateway-request-validator property.

Enable Basic Request Validation in API Gateway

See: newRequestValidator smart constructor.

Constructors

RequestValidator'

Fields

validateRequestParameters :: Maybe Bool

A Boolean flag to indicate whether to validate request parameters (true) or not (false).

name :: Maybe Text

The name of this RequestValidator

validateRequestBody :: Maybe Bool

A Boolean flag to indicate whether to validate a request body according to the configured Model schema.

id :: Maybe Text

The identifier of this RequestValidator.

newRequestValidator :: RequestValidator

Create a value of RequestValidator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validateRequestParameters:RequestValidator', requestValidator_validateRequestParameters - A Boolean flag to indicate whether to validate request parameters (true) or not (false).

$sel:name:RequestValidator', requestValidator_name - The name of this RequestValidator

$sel:validateRequestBody:RequestValidator', requestValidator_validateRequestBody - A Boolean flag to indicate whether to validate a request body according to the configured Model schema.

$sel:id:RequestValidator', requestValidator_id - The identifier of this RequestValidator.

Response Lenses

requestValidator_validateRequestParameters :: Lens' RequestValidator (Maybe Bool)

A Boolean flag to indicate whether to validate request parameters (true) or not (false).

requestValidator_name :: Lens' RequestValidator (Maybe Text)

The name of this RequestValidator

requestValidator_validateRequestBody :: Lens' RequestValidator (Maybe Bool)

A Boolean flag to indicate whether to validate a request body according to the configured Model schema.

requestValidator_id :: Lens' RequestValidator (Maybe Text)

The identifier of this RequestValidator.