Network.AWS.APIGateway.UpdateUsage

Description

Grants a temporary extension to the remaining quota of a usage plan associated with a specified API key.

Creating a Request

data UpdateUsage

The PATCH request to grant a temporary extension to the remaining quota of a usage plan associated with a specified API key.

See: newUpdateUsage smart constructor.

Constructors

UpdateUsage'

Fields

patchOperations :: Maybe [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

usagePlanId :: Text

Required

The Id of the usage plan associated with the usage data.

keyId :: Text

Required

The identifier of the API key associated with the usage plan in which a temporary extension is granted to the remaining quota.

newUpdateUsage

Arguments

:: Text
$sel:usagePlanId:UpdateUsage'

-> Text
$sel:keyId:UpdateUsage'

-> UpdateUsage

Create a value of UpdateUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:patchOperations:UpdateUsage', updateUsage_patchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

$sel:usagePlanId:UpdateUsage', updateUsage_usagePlanId - [Required] The Id of the usage plan associated with the usage data.

$sel:keyId:UpdateUsage', updateUsage_keyId - [Required] The identifier of the API key associated with the usage plan in which a temporary extension is granted to the remaining quota.

Request Lenses

updateUsage_patchOperations :: Lens' UpdateUsage (Maybe [PatchOperation])

A list of update operations to be applied to the specified resource and in the order specified in this list.

updateUsage_usagePlanId :: Lens' UpdateUsage Text

Required

The Id of the usage plan associated with the usage data.

updateUsage_keyId :: Lens' UpdateUsage Text

Required

The identifier of the API key associated with the usage plan in which a temporary extension is granted to the remaining quota.

Destructuring the Response

data Usage

Represents the usage data of a usage plan.

Create and Use Usage Plans, Manage Usage in a Usage Plan

See: newUsage smart constructor.

Constructors

Usage'

Fields

startDate :: Maybe Text

The starting date of the usage data.

items :: Maybe (HashMap Text [[Integer]])

The usage data, as daily logs of used and remaining quotas, over the specified time interval indexed over the API keys in a usage plan. For example, {..., "values" : { "{api_key}" : [ [0, 100], [10, 90], [100, 10]]}, where {api_key} stands for an API key value and the daily log entry is of the format [used quota, remaining quota].

position :: Maybe Text

endDate :: Maybe Text

The ending date of the usage data.

usagePlanId :: Maybe Text

The plan Id associated with this usage data.

newUsage :: Usage

Create a value of Usage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startDate:Usage', usage_startDate - The starting date of the usage data.

$sel:items:Usage', usage_items - The usage data, as daily logs of used and remaining quotas, over the specified time interval indexed over the API keys in a usage plan. For example, {..., "values" : { "{api_key}" : [ [0, 100], [10, 90], [100, 10]]}, where {api_key} stands for an API key value and the daily log entry is of the format [used quota, remaining quota].

$sel:position:Usage', usage_position - Undocumented member.

$sel:endDate:Usage', usage_endDate - The ending date of the usage data.

$sel:usagePlanId:Usage', usage_usagePlanId - The plan Id associated with this usage data.

Response Lenses

usage_startDate :: Lens' Usage (Maybe Text)

The starting date of the usage data.

usage_items :: Lens' Usage (Maybe (HashMap Text [[Integer]]))

The usage data, as daily logs of used and remaining quotas, over the specified time interval indexed over the API keys in a usage plan. For example, {..., "values" : { "{api_key}" : [ [0, 100], [10, 90], [100, 10]]}, where {api_key} stands for an API key value and the daily log entry is of the format [used quota, remaining quota].

usage_position :: Lens' Usage (Maybe Text)

Undocumented member.

usage_endDate :: Lens' Usage (Maybe Text)

The ending date of the usage data.

usage_usagePlanId :: Lens' Usage (Maybe Text)

The plan Id associated with this usage data.