Network.AWS.APIGateway.UpdateVpcLink

Description

Updates an existing VpcLink of a specified identifier.

Creating a Request

data UpdateVpcLink

Updates an existing VpcLink of a specified identifier.

See: newUpdateVpcLink smart constructor.

Constructors

UpdateVpcLink'

Fields

patchOperations :: Maybe [PatchOperation]

A list of update operations to be applied to the specified resource and in the order specified in this list.

vpcLinkId :: Text

Required

The identifier of the VpcLink. It is used in an Integration to reference this VpcLink.

newUpdateVpcLink

Arguments

:: Text
$sel:vpcLinkId:UpdateVpcLink'

-> UpdateVpcLink

Create a value of UpdateVpcLink with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:patchOperations:UpdateVpcLink', updateVpcLink_patchOperations - A list of update operations to be applied to the specified resource and in the order specified in this list.

$sel:vpcLinkId:UpdateVpcLink', updateVpcLink_vpcLinkId - [Required] The identifier of the VpcLink. It is used in an Integration to reference this VpcLink.

Request Lenses

updateVpcLink_patchOperations :: Lens' UpdateVpcLink (Maybe [PatchOperation])

A list of update operations to be applied to the specified resource and in the order specified in this list.

updateVpcLink_vpcLinkId :: Lens' UpdateVpcLink Text

Required

The identifier of the VpcLink. It is used in an Integration to reference this VpcLink.

Destructuring the Response

data VpcLink

An API Gateway VPC link for a RestApi to access resources in an Amazon Virtual Private Cloud (VPC).

To enable access to a resource in an Amazon Virtual Private Cloud through Amazon API Gateway, you, as an API developer, create a VpcLink resource targeted for one or more network load balancers of the VPC and then integrate an API method with a private integration that uses the VpcLink. The private integration has an integration type of HTTP or HTTP_PROXY and has a connection type of VPC_LINK. The integration uses the connectionId property to identify the VpcLink used.

See: newVpcLink smart constructor.

Constructors

VpcLink'

Fields

statusMessage :: Maybe Text

A description about the VPC link status.

status :: Maybe VpcLinkStatus

The status of the VPC link. The valid values are AVAILABLE, PENDING, DELETING, or FAILED. Deploying an API will wait if the status is PENDING and will fail if the status is DELETING.

id :: Maybe Text

The identifier of the VpcLink. It is used in an Integration to reference this VpcLink.

targetArns :: Maybe [Text]

The ARN of the network load balancer of the VPC targeted by the VPC link. The network load balancer must be owned by the same AWS account of the API owner.

name :: Maybe Text

The name used to label and identify the VPC link.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

description :: Maybe Text

The description of the VPC link.

newVpcLink :: VpcLink

Create a value of VpcLink with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:VpcLink', vpcLink_statusMessage - A description about the VPC link status.

$sel:status:VpcLink', vpcLink_status - The status of the VPC link. The valid values are AVAILABLE, PENDING, DELETING, or FAILED. Deploying an API will wait if the status is PENDING and will fail if the status is DELETING.

$sel:id:VpcLink', vpcLink_id - The identifier of the VpcLink. It is used in an Integration to reference this VpcLink.

$sel:targetArns:VpcLink', vpcLink_targetArns - The ARN of the network load balancer of the VPC targeted by the VPC link. The network load balancer must be owned by the same AWS account of the API owner.

$sel:name:VpcLink', vpcLink_name - The name used to label and identify the VPC link.

$sel:tags:VpcLink', vpcLink_tags - The collection of tags. Each tag element is associated with a given resource.

$sel:description:VpcLink', vpcLink_description - The description of the VPC link.

Response Lenses

vpcLink_statusMessage :: Lens' VpcLink (Maybe Text)

A description about the VPC link status.

vpcLink_status :: Lens' VpcLink (Maybe VpcLinkStatus)

The status of the VPC link. The valid values are AVAILABLE, PENDING, DELETING, or FAILED. Deploying an API will wait if the status is PENDING and will fail if the status is DELETING.

vpcLink_id :: Lens' VpcLink (Maybe Text)

The identifier of the VpcLink. It is used in an Integration to reference this VpcLink.

vpcLink_targetArns :: Lens' VpcLink (Maybe [Text])

The ARN of the network load balancer of the VPC targeted by the VPC link. The network load balancer must be owned by the same AWS account of the API owner.

vpcLink_name :: Lens' VpcLink (Maybe Text)

The name used to label and identify the VPC link.

vpcLink_tags :: Lens' VpcLink (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.

vpcLink_description :: Lens' VpcLink (Maybe Text)

The description of the VPC link.