Network.AWS.APIGatewayManagementAPI.Lens

Description

Operations

DeleteConnection

deleteConnection_connectionId :: Lens' DeleteConnection Text

Undocumented member.

GetConnection

getConnection_connectionId :: Lens' GetConnection Text

Undocumented member.

getConnectionResponse_connectedAt :: Lens' GetConnectionResponse (Maybe UTCTime)

The time in ISO 8601 format for when the connection was established.

getConnectionResponse_lastActiveAt :: Lens' GetConnectionResponse (Maybe UTCTime)

The time in ISO 8601 format for when the connection was last active.

getConnectionResponse_identity :: Lens' GetConnectionResponse (Maybe Identity)

Undocumented member.

getConnectionResponse_httpStatus :: Lens' GetConnectionResponse Int

The response's http status code.

PostToConnection

postToConnection_connectionId :: Lens' PostToConnection Text

The identifier of the connection that a specific client is using.

postToConnection_data :: Lens' PostToConnection ByteString

The data to be sent to the client specified by its connection id.

Types

Identity

identity_sourceIp :: Lens' Identity Text

The source IP address of the TCP connection making the request to API Gateway.

identity_userAgent :: Lens' Identity Text

The User Agent of the API caller.