Network.AWS.APIGatewayManagementAPI.Types

Description

Service Configuration

defaultService :: Service

API version 2018-11-29 of the Amazon ApiGatewayManagementApi SDK configuration.

Errors

_PayloadTooLargeException :: AsError a => Getting (First ServiceError) a ServiceError

The data has exceeded the maximum size allowed.

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

The caller is not authorized to invoke this operation.

_GoneException :: AsError a => Getting (First ServiceError) a ServiceError

The connection with the provided id no longer exists.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The client is sending more than the allowed number of requests per unit of time or the WebSocket client side buffer is full.

Identity

data Identity

See: newIdentity smart constructor.

Constructors

Identity'

Fields

sourceIp :: Text

The source IP address of the TCP connection making the request to API Gateway.

userAgent :: Text

The User Agent of the API caller.

newIdentity

Arguments

:: Text
$sel:sourceIp:Identity'

-> Text
$sel:userAgent:Identity'

-> Identity

Create a value of Identity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceIp:Identity', identity_sourceIp - The source IP address of the TCP connection making the request to API Gateway.

$sel:userAgent:Identity', identity_userAgent - The User Agent of the API caller.

identity_sourceIp :: Lens' Identity Text

The source IP address of the TCP connection making the request to API Gateway.

identity_userAgent :: Lens' Identity Text

The User Agent of the API caller.