Network.AWS.AWSHealth.DescribeEntityAggregates

Description

Returns the number of entities that are affected by each of the specified events. If no events are specified, the counts of all affected entities are returned.

Creating a Request

data DescribeEntityAggregates

See: newDescribeEntityAggregates smart constructor.

Constructors

DescribeEntityAggregates'

Fields

eventArns :: Maybe (NonEmpty Text)

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

newDescribeEntityAggregates :: DescribeEntityAggregates

Create a value of DescribeEntityAggregates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventArns:DescribeEntityAggregates', describeEntityAggregates_eventArns - A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

Request Lenses

describeEntityAggregates_eventArns :: Lens' DescribeEntityAggregates (Maybe (NonEmpty Text))

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

Destructuring the Response

data DescribeEntityAggregatesResponse

See: newDescribeEntityAggregatesResponse smart constructor.

Constructors

DescribeEntityAggregatesResponse'

Fields

entityAggregates :: Maybe [EntityAggregate]

The number of entities that are affected by each of the specified events.

httpStatus :: Int

The response's http status code.

newDescribeEntityAggregatesResponse

Arguments

:: Int
$sel:httpStatus:DescribeEntityAggregatesResponse'

-> DescribeEntityAggregatesResponse

Create a value of DescribeEntityAggregatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entityAggregates:DescribeEntityAggregatesResponse', describeEntityAggregatesResponse_entityAggregates - The number of entities that are affected by each of the specified events.

$sel:httpStatus:DescribeEntityAggregatesResponse', describeEntityAggregatesResponse_httpStatus - The response's http status code.

Response Lenses

describeEntityAggregatesResponse_entityAggregates :: Lens' DescribeEntityAggregatesResponse (Maybe [EntityAggregate])

The number of entities that are affected by each of the specified events.

describeEntityAggregatesResponse_httpStatus :: Lens' DescribeEntityAggregatesResponse Int

The response's http status code.