Network.AWS.AWSHealth.DescribeEventAggregates

Description

Returns the number of events of each event type (issue, scheduled change, and account notification). If no filter is specified, the counts of all events in each category are returned.

This API operation uses pagination. Specify the nextToken parameter in the next request to return more results.

This operation returns paginated results.

Creating a Request

data DescribeEventAggregates

See: newDescribeEventAggregates smart constructor.

Constructors

DescribeEventAggregates'

Fields

nextToken :: Maybe Text

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

filter' :: Maybe EventFilter

Values to narrow the results returned.

maxResults :: Maybe Natural

The maximum number of items to return in one batch, between 10 and 100, inclusive.

aggregateField :: EventAggregateField

The only currently supported value is eventTypeCategory.

newDescribeEventAggregates

Arguments

:: EventAggregateField
$sel:aggregateField:DescribeEventAggregates'

-> DescribeEventAggregates

Create a value of DescribeEventAggregates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeEventAggregates', describeEventAggregates_nextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

$sel:filter':DescribeEventAggregates', describeEventAggregates_filter - Values to narrow the results returned.

$sel:maxResults:DescribeEventAggregates', describeEventAggregates_maxResults - The maximum number of items to return in one batch, between 10 and 100, inclusive.

$sel:aggregateField:DescribeEventAggregates', describeEventAggregates_aggregateField - The only currently supported value is eventTypeCategory.

Request Lenses

describeEventAggregates_nextToken :: Lens' DescribeEventAggregates (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeEventAggregates_filter :: Lens' DescribeEventAggregates (Maybe EventFilter)

Values to narrow the results returned.

describeEventAggregates_maxResults :: Lens' DescribeEventAggregates (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

describeEventAggregates_aggregateField :: Lens' DescribeEventAggregates EventAggregateField

The only currently supported value is eventTypeCategory.

Destructuring the Response

data DescribeEventAggregatesResponse

See: newDescribeEventAggregatesResponse smart constructor.

Constructors

DescribeEventAggregatesResponse'

Fields

nextToken :: Maybe Text

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

eventAggregates :: Maybe [EventAggregate]

The number of events in each category that meet the optional filter criteria.

httpStatus :: Int

The response's http status code.

newDescribeEventAggregatesResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventAggregatesResponse'

-> DescribeEventAggregatesResponse

Create a value of DescribeEventAggregatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeEventAggregates', describeEventAggregatesResponse_nextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

$sel:eventAggregates:DescribeEventAggregatesResponse', describeEventAggregatesResponse_eventAggregates - The number of events in each category that meet the optional filter criteria.

$sel:httpStatus:DescribeEventAggregatesResponse', describeEventAggregatesResponse_httpStatus - The response's http status code.

Response Lenses

describeEventAggregatesResponse_nextToken :: Lens' DescribeEventAggregatesResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeEventAggregatesResponse_eventAggregates :: Lens' DescribeEventAggregatesResponse (Maybe [EventAggregate])

The number of events in each category that meet the optional filter criteria.

describeEventAggregatesResponse_httpStatus :: Lens' DescribeEventAggregatesResponse Int

The response's http status code.