Network.AWS.AWSHealth.DescribeEventDetails

Description

Returns detailed information about one or more specified events. Information includes standard event data (AWS Region, service, and so on, as returned by DescribeEvents), a detailed event description, and possible additional metadata that depends upon the nature of the event. Affected entities are not included. To retrieve the entities, use the DescribeAffectedEntities operation.

If a specified event can't be retrieved, an error message is returned for that event.

This operation supports resource-level permissions. You can use this operation to allow or deny access to specific AWS Health events. For more information, see Resource- and action-based conditions in the AWS Health User Guide.

Creating a Request

data DescribeEventDetails

See: newDescribeEventDetails smart constructor.

Constructors

DescribeEventDetails'

Fields

locale :: Maybe Text

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

eventArns :: NonEmpty Text

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

newDescribeEventDetails

Arguments

:: NonEmpty Text
$sel:eventArns:DescribeEventDetails'

-> DescribeEventDetails

Create a value of DescribeEventDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locale:DescribeEventDetails', describeEventDetails_locale - The locale (language) to return information in. English (en) is the default and the only supported value at this time.

$sel:eventArns:DescribeEventDetails', describeEventDetails_eventArns - A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

Request Lenses

describeEventDetails_locale :: Lens' DescribeEventDetails (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

describeEventDetails_eventArns :: Lens' DescribeEventDetails (NonEmpty Text)

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

Destructuring the Response

data DescribeEventDetailsResponse

See: newDescribeEventDetailsResponse smart constructor.

Constructors

DescribeEventDetailsResponse'

Fields

successfulSet :: Maybe [EventDetails]

Information about the events that could be retrieved.

failedSet :: Maybe [EventDetailsErrorItem]

Error messages for any events that could not be retrieved.

httpStatus :: Int

The response's http status code.

newDescribeEventDetailsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventDetailsResponse'

-> DescribeEventDetailsResponse

Create a value of DescribeEventDetailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:successfulSet:DescribeEventDetailsResponse', describeEventDetailsResponse_successfulSet - Information about the events that could be retrieved.

$sel:failedSet:DescribeEventDetailsResponse', describeEventDetailsResponse_failedSet - Error messages for any events that could not be retrieved.

$sel:httpStatus:DescribeEventDetailsResponse', describeEventDetailsResponse_httpStatus - The response's http status code.

Response Lenses

describeEventDetailsResponse_successfulSet :: Lens' DescribeEventDetailsResponse (Maybe [EventDetails])

Information about the events that could be retrieved.

describeEventDetailsResponse_failedSet :: Lens' DescribeEventDetailsResponse (Maybe [EventDetailsErrorItem])

Error messages for any events that could not be retrieved.

describeEventDetailsResponse_httpStatus :: Lens' DescribeEventDetailsResponse Int

The response's http status code.