Network.AWS.AWSHealth.DescribeEventTypes

Description

Returns the event types that meet the specified filter criteria. You can use this API operation to find information about the AWS Health event, such as the category, AWS service, and event code. The metadata for each event appears in the EventType object.

If you don't specify a filter criteria, the API operation returns all event types, in no particular order.

This API operation uses pagination. Specify the nextToken parameter in the next request to return more results.

This operation returns paginated results.

Creating a Request

data DescribeEventTypes

See: newDescribeEventTypes smart constructor.

Constructors

DescribeEventTypes'

Fields

locale :: Maybe Text

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

nextToken :: Maybe Text

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

filter' :: Maybe EventTypeFilter

Values to narrow the results returned.

maxResults :: Maybe Natural

The maximum number of items to return in one batch, between 10 and 100, inclusive.

newDescribeEventTypes :: DescribeEventTypes

Create a value of DescribeEventTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locale:DescribeEventTypes', describeEventTypes_locale - The locale (language) to return information in. English (en) is the default and the only supported value at this time.

$sel:nextToken:DescribeEventTypes', describeEventTypes_nextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

$sel:filter':DescribeEventTypes', describeEventTypes_filter - Values to narrow the results returned.

$sel:maxResults:DescribeEventTypes', describeEventTypes_maxResults - The maximum number of items to return in one batch, between 10 and 100, inclusive.

Request Lenses

describeEventTypes_locale :: Lens' DescribeEventTypes (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

describeEventTypes_nextToken :: Lens' DescribeEventTypes (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeEventTypes_filter :: Lens' DescribeEventTypes (Maybe EventTypeFilter)

Values to narrow the results returned.

describeEventTypes_maxResults :: Lens' DescribeEventTypes (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

Destructuring the Response

data DescribeEventTypesResponse

See: newDescribeEventTypesResponse smart constructor.

Constructors

DescribeEventTypesResponse'

Fields

eventTypes :: Maybe [EventType]

A list of event types that match the filter criteria. Event types have a category (issue, accountNotification, or scheduledChange), a service (for example, EC2, RDS, DATAPIPELINE, BILLING), and a code (in the format AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT).

nextToken :: Maybe Text

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

httpStatus :: Int

The response's http status code.

newDescribeEventTypesResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventTypesResponse'

-> DescribeEventTypesResponse

Create a value of DescribeEventTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventTypes:DescribeEventTypesResponse', describeEventTypesResponse_eventTypes - A list of event types that match the filter criteria. Event types have a category (issue, accountNotification, or scheduledChange), a service (for example, EC2, RDS, DATAPIPELINE, BILLING), and a code (in the format AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT).

$sel:nextToken:DescribeEventTypes', describeEventTypesResponse_nextToken - If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

$sel:httpStatus:DescribeEventTypesResponse', describeEventTypesResponse_httpStatus - The response's http status code.

Response Lenses

describeEventTypesResponse_eventTypes :: Lens' DescribeEventTypesResponse (Maybe [EventType])

A list of event types that match the filter criteria. Event types have a category (issue, accountNotification, or scheduledChange), a service (for example, EC2, RDS, DATAPIPELINE, BILLING), and a code (in the format AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT).

describeEventTypesResponse_nextToken :: Lens' DescribeEventTypesResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeEventTypesResponse_httpStatus :: Lens' DescribeEventTypesResponse Int

The response's http status code.