Network.AWS.AWSHealth.DisableHealthServiceAccessForOrganization

Description

Disables AWS Health from working with AWS Organizations. To call this operation, you must sign in as an AWS Identity and Access Management (IAM) user, assume an IAM role, or sign in as the root user (not recommended) in the organization's management account. For more information, see Aggregating AWS Health events in the AWS Health User Guide.

This operation doesn't remove the service-linked role from the management account in your organization. You must use the IAM console, API, or AWS Command Line Interface (AWS CLI) to remove the service-linked role. For more information, see Deleting a Service-Linked Role in the IAM User Guide.

You can also disable the organizational feature by using the Organizations DisableAWSServiceAccess API operation. After you call this operation, AWS Health stops aggregating events for all other AWS accounts in your organization. If you call the AWS Health API operations for organizational view, AWS Health returns an error. AWS Health continues to aggregate health events for your AWS account.

Creating a Request

data DisableHealthServiceAccessForOrganization

See: newDisableHealthServiceAccessForOrganization smart constructor.

Constructors

DisableHealthServiceAccessForOrganization'

newDisableHealthServiceAccessForOrganization :: DisableHealthServiceAccessForOrganization

Create a value of DisableHealthServiceAccessForOrganization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DisableHealthServiceAccessForOrganizationResponse

See: newDisableHealthServiceAccessForOrganizationResponse smart constructor.

Constructors

DisableHealthServiceAccessForOrganizationResponse'

newDisableHealthServiceAccessForOrganizationResponse :: DisableHealthServiceAccessForOrganizationResponse

Create a value of DisableHealthServiceAccessForOrganizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.