Network.AWS.AWSHealth.Lens

Description

Operations

EnableHealthServiceAccessForOrganization

DescribeEntityAggregates

describeEntityAggregates_eventArns :: Lens' DescribeEntityAggregates (Maybe (NonEmpty Text))

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

describeEntityAggregatesResponse_entityAggregates :: Lens' DescribeEntityAggregatesResponse (Maybe [EntityAggregate])

The number of entities that are affected by each of the specified events.

describeEntityAggregatesResponse_httpStatus :: Lens' DescribeEntityAggregatesResponse Int

The response's http status code.

DescribeEvents

describeEvents_locale :: Lens' DescribeEvents (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

describeEvents_nextToken :: Lens' DescribeEvents (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeEvents_filter :: Lens' DescribeEvents (Maybe EventFilter)

Values to narrow the results returned.

describeEvents_maxResults :: Lens' DescribeEvents (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

describeEventsResponse_nextToken :: Lens' DescribeEventsResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeEventsResponse_events :: Lens' DescribeEventsResponse (Maybe [Event])

The events that match the specified filter criteria.

describeEventsResponse_httpStatus :: Lens' DescribeEventsResponse Int

The response's http status code.

DescribeEventsForOrganization

describeEventsForOrganization_locale :: Lens' DescribeEventsForOrganization (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

describeEventsForOrganization_nextToken :: Lens' DescribeEventsForOrganization (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeEventsForOrganization_filter :: Lens' DescribeEventsForOrganization (Maybe OrganizationEventFilter)

Values to narrow the results returned.

describeEventsForOrganization_maxResults :: Lens' DescribeEventsForOrganization (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

describeEventsForOrganizationResponse_nextToken :: Lens' DescribeEventsForOrganizationResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeEventsForOrganizationResponse_events :: Lens' DescribeEventsForOrganizationResponse (Maybe [OrganizationEvent])

The events that match the specified filter criteria.

describeEventsForOrganizationResponse_httpStatus :: Lens' DescribeEventsForOrganizationResponse Int

The response's http status code.

DescribeAffectedAccountsForOrganization

describeAffectedAccountsForOrganization_nextToken :: Lens' DescribeAffectedAccountsForOrganization (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeAffectedAccountsForOrganization_maxResults :: Lens' DescribeAffectedAccountsForOrganization (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

describeAffectedAccountsForOrganization_eventArn :: Lens' DescribeAffectedAccountsForOrganization Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

describeAffectedAccountsForOrganizationResponse_affectedAccounts :: Lens' DescribeAffectedAccountsForOrganizationResponse (Maybe [Text])

A JSON set of elements of the affected accounts.

describeAffectedAccountsForOrganizationResponse_eventScopeCode :: Lens' DescribeAffectedAccountsForOrganizationResponse (Maybe EventScopeCode)

This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

describeAffectedAccountsForOrganizationResponse_nextToken :: Lens' DescribeAffectedAccountsForOrganizationResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeAffectedAccountsForOrganizationResponse_httpStatus :: Lens' DescribeAffectedAccountsForOrganizationResponse Int

The response's http status code.

DescribeEventDetails

describeEventDetails_locale :: Lens' DescribeEventDetails (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

describeEventDetails_eventArns :: Lens' DescribeEventDetails (NonEmpty Text)

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

describeEventDetailsResponse_successfulSet :: Lens' DescribeEventDetailsResponse (Maybe [EventDetails])

Information about the events that could be retrieved.

describeEventDetailsResponse_failedSet :: Lens' DescribeEventDetailsResponse (Maybe [EventDetailsErrorItem])

Error messages for any events that could not be retrieved.

describeEventDetailsResponse_httpStatus :: Lens' DescribeEventDetailsResponse Int

The response's http status code.

DescribeEventAggregates

describeEventAggregates_nextToken :: Lens' DescribeEventAggregates (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeEventAggregates_filter :: Lens' DescribeEventAggregates (Maybe EventFilter)

Values to narrow the results returned.

describeEventAggregates_maxResults :: Lens' DescribeEventAggregates (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

describeEventAggregates_aggregateField :: Lens' DescribeEventAggregates EventAggregateField

The only currently supported value is eventTypeCategory.

describeEventAggregatesResponse_nextToken :: Lens' DescribeEventAggregatesResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeEventAggregatesResponse_eventAggregates :: Lens' DescribeEventAggregatesResponse (Maybe [EventAggregate])

The number of events in each category that meet the optional filter criteria.

describeEventAggregatesResponse_httpStatus :: Lens' DescribeEventAggregatesResponse Int

The response's http status code.

DescribeAffectedEntities

describeAffectedEntities_locale :: Lens' DescribeAffectedEntities (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

describeAffectedEntities_nextToken :: Lens' DescribeAffectedEntities (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeAffectedEntities_maxResults :: Lens' DescribeAffectedEntities (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

describeAffectedEntities_filter :: Lens' DescribeAffectedEntities EntityFilter

Values to narrow the results returned. At least one event ARN is required.

describeAffectedEntitiesResponse_entities :: Lens' DescribeAffectedEntitiesResponse (Maybe [AffectedEntity])

The entities that match the filter criteria.

describeAffectedEntitiesResponse_nextToken :: Lens' DescribeAffectedEntitiesResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeAffectedEntitiesResponse_httpStatus :: Lens' DescribeAffectedEntitiesResponse Int

The response's http status code.

DescribeEventTypes

describeEventTypes_locale :: Lens' DescribeEventTypes (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

describeEventTypes_nextToken :: Lens' DescribeEventTypes (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeEventTypes_filter :: Lens' DescribeEventTypes (Maybe EventTypeFilter)

Values to narrow the results returned.

describeEventTypes_maxResults :: Lens' DescribeEventTypes (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

describeEventTypesResponse_eventTypes :: Lens' DescribeEventTypesResponse (Maybe [EventType])

A list of event types that match the filter criteria. Event types have a category (issue, accountNotification, or scheduledChange), a service (for example, EC2, RDS, DATAPIPELINE, BILLING), and a code (in the format AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT).

describeEventTypesResponse_nextToken :: Lens' DescribeEventTypesResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeEventTypesResponse_httpStatus :: Lens' DescribeEventTypesResponse Int

The response's http status code.

DescribeAffectedEntitiesForOrganization

describeAffectedEntitiesForOrganization_locale :: Lens' DescribeAffectedEntitiesForOrganization (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

describeAffectedEntitiesForOrganization_nextToken :: Lens' DescribeAffectedEntitiesForOrganization (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeAffectedEntitiesForOrganization_maxResults :: Lens' DescribeAffectedEntitiesForOrganization (Maybe Natural)

The maximum number of items to return in one batch, between 10 and 100, inclusive.

describeAffectedEntitiesForOrganization_organizationEntityFilters :: Lens' DescribeAffectedEntitiesForOrganization (NonEmpty EventAccountFilter)

A JSON set of elements including the awsAccountId and the eventArn.

describeAffectedEntitiesForOrganizationResponse_entities :: Lens' DescribeAffectedEntitiesForOrganizationResponse (Maybe [AffectedEntity])

A JSON set of elements including the awsAccountId and its entityArn, entityValue and its entityArn, lastUpdatedTime, and statusCode.

describeAffectedEntitiesForOrganizationResponse_failedSet :: Lens' DescribeAffectedEntitiesForOrganizationResponse (Maybe [OrganizationAffectedEntitiesErrorItem])

A JSON set of elements of the failed response, including the awsAccountId, errorMessage, errorName, and eventArn.

describeAffectedEntitiesForOrganizationResponse_nextToken :: Lens' DescribeAffectedEntitiesForOrganizationResponse (Maybe Text)

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

describeAffectedEntitiesForOrganizationResponse_httpStatus :: Lens' DescribeAffectedEntitiesForOrganizationResponse Int

The response's http status code.

DescribeHealthServiceStatusForOrganization

describeHealthServiceStatusForOrganizationResponse_healthServiceAccessStatusForOrganization :: Lens' DescribeHealthServiceStatusForOrganizationResponse (Maybe Text)

Information about the status of enabling or disabling AWS Health Organizational View in your organization.

Valid values are ENABLED | DISABLED | PENDING.

describeHealthServiceStatusForOrganizationResponse_httpStatus :: Lens' DescribeHealthServiceStatusForOrganizationResponse Int

The response's http status code.

DescribeEventDetailsForOrganization

describeEventDetailsForOrganization_locale :: Lens' DescribeEventDetailsForOrganization (Maybe Text)

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

describeEventDetailsForOrganization_organizationEventDetailFilters :: Lens' DescribeEventDetailsForOrganization (NonEmpty EventAccountFilter)

A set of JSON elements that includes the awsAccountId and the eventArn.

describeEventDetailsForOrganizationResponse_successfulSet :: Lens' DescribeEventDetailsForOrganizationResponse (Maybe [OrganizationEventDetails])

Information about the events that could be retrieved.

describeEventDetailsForOrganizationResponse_failedSet :: Lens' DescribeEventDetailsForOrganizationResponse (Maybe [OrganizationEventDetailsErrorItem])

Error messages for any events that could not be retrieved.

describeEventDetailsForOrganizationResponse_httpStatus :: Lens' DescribeEventDetailsForOrganizationResponse Int

The response's http status code.

DisableHealthServiceAccessForOrganization

Types

AffectedEntity

affectedEntity_lastUpdatedTime :: Lens' AffectedEntity (Maybe UTCTime)

The most recent time that the entity was updated.

affectedEntity_entityValue :: Lens' AffectedEntity (Maybe Text)

The ID of the affected entity.

affectedEntity_entityUrl :: Lens' AffectedEntity (Maybe Text)

The URL of the affected entity.

affectedEntity_awsAccountId :: Lens' AffectedEntity (Maybe Text)

The 12-digit AWS account number that contains the affected entity.

affectedEntity_eventArn :: Lens' AffectedEntity (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

affectedEntity_entityArn :: Lens' AffectedEntity (Maybe Text)

The unique identifier for the entity. Format: arn:aws:health:entity-region:aws-account:entity/entity-id . Example: arn:aws:health:us-east-1:111222333444:entity/AVh5GGT7ul1arKr1sE1K

affectedEntity_tags :: Lens' AffectedEntity (Maybe (HashMap Text Text))

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

affectedEntity_statusCode :: Lens' AffectedEntity (Maybe EntityStatusCode)

The most recent status of the entity affected by the event. The possible values are IMPAIRED, UNIMPAIRED, and UNKNOWN.

DateTimeRange

dateTimeRange_to :: Lens' DateTimeRange (Maybe UTCTime)

The ending date and time of a time range.

dateTimeRange_from :: Lens' DateTimeRange (Maybe UTCTime)

The starting date and time of a time range.

EntityAggregate

entityAggregate_count :: Lens' EntityAggregate (Maybe Int)

The number of entities that match the criteria for the specified events.

entityAggregate_eventArn :: Lens' EntityAggregate (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

EntityFilter

entityFilter_statusCodes :: Lens' EntityFilter (Maybe (NonEmpty EntityStatusCode))

A list of entity status codes (IMPAIRED, UNIMPAIRED, or UNKNOWN).

entityFilter_entityArns :: Lens' EntityFilter (Maybe (NonEmpty Text))

A list of entity ARNs (unique identifiers).

entityFilter_entityValues :: Lens' EntityFilter (Maybe (NonEmpty Text))

A list of IDs for affected entities.

entityFilter_tags :: Lens' EntityFilter (Maybe [HashMap Text Text])

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

entityFilter_lastUpdatedTimes :: Lens' EntityFilter (Maybe (NonEmpty DateTimeRange))

A list of the most recent dates and times that the entity was updated.

entityFilter_eventArns :: Lens' EntityFilter (NonEmpty Text)

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

Event

event_lastUpdatedTime :: Lens' Event (Maybe UTCTime)

The most recent date and time that the event was updated.

event_arn :: Lens' Event (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

event_service :: Lens' Event (Maybe Text)

The AWS service that is affected by the event. For example, EC2, RDS.

event_startTime :: Lens' Event (Maybe UTCTime)

The date and time that the event began.

event_eventScopeCode :: Lens' Event (Maybe EventScopeCode)

This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

event_eventTypeCode :: Lens' Event (Maybe Text)

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

event_eventTypeCategory :: Lens' Event (Maybe EventTypeCategory)

The category of the event. Possible values are issue, scheduledChange, and accountNotification.

event_availabilityZone :: Lens' Event (Maybe Text)

The AWS Availability Zone of the event. For example, us-east-1a.

event_endTime :: Lens' Event (Maybe UTCTime)

The date and time that the event ended.

event_region :: Lens' Event (Maybe Text)

The AWS Region name of the event.

event_statusCode :: Lens' Event (Maybe EventStatusCode)

The most recent status of the event. Possible values are open, closed, and upcoming.

EventAccountFilter

eventAccountFilter_awsAccountId :: Lens' EventAccountFilter (Maybe Text)

The 12-digit AWS account numbers that contains the affected entities.

eventAccountFilter_eventArn :: Lens' EventAccountFilter Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

EventAggregate

eventAggregate_count :: Lens' EventAggregate (Maybe Int)

The number of events of the associated issue type.

eventAggregate_aggregateValue :: Lens' EventAggregate (Maybe Text)

The issue type for the associated count.

EventDescription

eventDescription_latestDescription :: Lens' EventDescription (Maybe Text)

The most recent description of the event.

EventDetails

eventDetails_event :: Lens' EventDetails (Maybe Event)

Summary information about the event.

eventDetails_eventDescription :: Lens' EventDetails (Maybe EventDescription)

The most recent description of the event.

eventDetails_eventMetadata :: Lens' EventDetails (Maybe (HashMap Text Text))

Additional metadata about the event.

EventDetailsErrorItem

eventDetailsErrorItem_eventArn :: Lens' EventDetailsErrorItem (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

eventDetailsErrorItem_errorName :: Lens' EventDetailsErrorItem (Maybe Text)

The name of the error.

eventDetailsErrorItem_errorMessage :: Lens' EventDetailsErrorItem (Maybe Text)

A message that describes the error.

EventFilter

eventFilter_eventArns :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

eventFilter_eventTypeCategories :: Lens' EventFilter (Maybe (NonEmpty EventTypeCategory))

A list of event type category codes (issue, scheduledChange, or accountNotification).

eventFilter_eventTypeCodes :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of unique identifiers for event types. For example, "AWS_EC2_SYSTEM_MAINTENANCE_EVENT","AWS_RDS_MAINTENANCE_SCHEDULED".

eventFilter_regions :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of AWS Regions.

eventFilter_eventStatusCodes :: Lens' EventFilter (Maybe (NonEmpty EventStatusCode))

A list of event status codes.

eventFilter_endTimes :: Lens' EventFilter (Maybe (NonEmpty DateTimeRange))

A list of dates and times that the event ended.

eventFilter_availabilityZones :: Lens' EventFilter (Maybe [Text])

A list of AWS Availability Zones.

eventFilter_entityArns :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of entity ARNs (unique identifiers).

eventFilter_entityValues :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of entity identifiers, such as EC2 instance IDs (i-34ab692e) or EBS volumes (vol-426ab23e).

eventFilter_startTimes :: Lens' EventFilter (Maybe (NonEmpty DateTimeRange))

A list of dates and times that the event began.

eventFilter_services :: Lens' EventFilter (Maybe (NonEmpty Text))

The AWS services associated with the event. For example, EC2, RDS.

eventFilter_tags :: Lens' EventFilter (Maybe [HashMap Text Text])

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

eventFilter_lastUpdatedTimes :: Lens' EventFilter (Maybe (NonEmpty DateTimeRange))

A list of dates and times that the event was last updated.

EventType

eventType_service :: Lens' EventType (Maybe Text)

The AWS service that is affected by the event. For example, EC2, RDS.

eventType_category :: Lens' EventType (Maybe EventTypeCategory)

A list of event type category codes (issue, scheduledChange, or accountNotification).

eventType_code :: Lens' EventType (Maybe Text)

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

EventTypeFilter

eventTypeFilter_eventTypeCategories :: Lens' EventTypeFilter (Maybe (NonEmpty EventTypeCategory))

A list of event type category codes (issue, scheduledChange, or accountNotification).

eventTypeFilter_eventTypeCodes :: Lens' EventTypeFilter (Maybe (NonEmpty Text))

A list of event type codes.

eventTypeFilter_services :: Lens' EventTypeFilter (Maybe (NonEmpty Text))

The AWS services associated with the event. For example, EC2, RDS.

OrganizationAffectedEntitiesErrorItem

organizationAffectedEntitiesErrorItem_awsAccountId :: Lens' OrganizationAffectedEntitiesErrorItem (Maybe Text)

The 12-digit AWS account numbers that contains the affected entities.

organizationAffectedEntitiesErrorItem_eventArn :: Lens' OrganizationAffectedEntitiesErrorItem (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

organizationAffectedEntitiesErrorItem_errorName :: Lens' OrganizationAffectedEntitiesErrorItem (Maybe Text)

The name of the error.

organizationAffectedEntitiesErrorItem_errorMessage :: Lens' OrganizationAffectedEntitiesErrorItem (Maybe Text)

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

OrganizationEvent

organizationEvent_lastUpdatedTime :: Lens' OrganizationEvent (Maybe UTCTime)

The most recent date and time that the event was updated.

organizationEvent_arn :: Lens' OrganizationEvent (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

organizationEvent_service :: Lens' OrganizationEvent (Maybe Text)

The AWS service that is affected by the event, such as EC2 and RDS.

organizationEvent_startTime :: Lens' OrganizationEvent (Maybe UTCTime)

The date and time that the event began.

organizationEvent_eventScopeCode :: Lens' OrganizationEvent (Maybe EventScopeCode)

This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

organizationEvent_eventTypeCode :: Lens' OrganizationEvent (Maybe Text)

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

organizationEvent_eventTypeCategory :: Lens' OrganizationEvent (Maybe EventTypeCategory)

The category of the event type.

organizationEvent_endTime :: Lens' OrganizationEvent (Maybe UTCTime)

The date and time that the event ended.

organizationEvent_region :: Lens' OrganizationEvent (Maybe Text)

The AWS Region name of the event.

organizationEvent_statusCode :: Lens' OrganizationEvent (Maybe EventStatusCode)

The most recent status of the event. Possible values are open, closed, and upcoming.

OrganizationEventDetails

organizationEventDetails_event :: Lens' OrganizationEventDetails (Maybe Event)

Undocumented member.

organizationEventDetails_eventDescription :: Lens' OrganizationEventDetails (Maybe EventDescription)

Undocumented member.

organizationEventDetails_awsAccountId :: Lens' OrganizationEventDetails (Maybe Text)

The 12-digit AWS account numbers that contains the affected entities.

organizationEventDetails_eventMetadata :: Lens' OrganizationEventDetails (Maybe (HashMap Text Text))

Additional metadata about the event.

OrganizationEventDetailsErrorItem

organizationEventDetailsErrorItem_awsAccountId :: Lens' OrganizationEventDetailsErrorItem (Maybe Text)

Error information returned when a DescribeEventDetailsForOrganization operation can't find a specified event.

organizationEventDetailsErrorItem_eventArn :: Lens' OrganizationEventDetailsErrorItem (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

organizationEventDetailsErrorItem_errorName :: Lens' OrganizationEventDetailsErrorItem (Maybe Text)

The name of the error.

organizationEventDetailsErrorItem_errorMessage :: Lens' OrganizationEventDetailsErrorItem (Maybe Text)

A message that describes the error.

If you call the DescribeEventDetailsForOrganization operation and receive one of the following errors, follow the recommendations in the message:

We couldn't find a public event that matches your request. To find an event that is account specific, you must enter an AWS account ID in the request.

We couldn't find an account specific event for the specified AWS account. To find an event that is public, you must enter a null value for the AWS account ID in the request.

Your AWS account doesn't include the AWS Support plan required to use the AWS Health API. You must have either a Business or Enterprise Support plan.

OrganizationEventFilter

organizationEventFilter_lastUpdatedTime :: Lens' OrganizationEventFilter (Maybe DateTimeRange)

Undocumented member.

organizationEventFilter_awsAccountIds :: Lens' OrganizationEventFilter (Maybe (NonEmpty Text))

A list of 12-digit AWS account numbers that contains the affected entities.

organizationEventFilter_eventTypeCategories :: Lens' OrganizationEventFilter (Maybe (NonEmpty EventTypeCategory))

A list of event type category codes (issue, scheduledChange, or accountNotification).

organizationEventFilter_eventTypeCodes :: Lens' OrganizationEventFilter (Maybe (NonEmpty Text))

A list of unique identifiers for event types. For example, "AWS_EC2_SYSTEM_MAINTENANCE_EVENT","AWS_RDS_MAINTENANCE_SCHEDULED".

organizationEventFilter_startTime :: Lens' OrganizationEventFilter (Maybe DateTimeRange)

Undocumented member.

organizationEventFilter_regions :: Lens' OrganizationEventFilter (Maybe (NonEmpty Text))

A list of AWS Regions.

organizationEventFilter_eventStatusCodes :: Lens' OrganizationEventFilter (Maybe (NonEmpty EventStatusCode))

A list of event status codes.

organizationEventFilter_endTime :: Lens' OrganizationEventFilter (Maybe DateTimeRange)

Undocumented member.

organizationEventFilter_entityArns :: Lens' OrganizationEventFilter (Maybe (NonEmpty Text))

A list of entity ARNs (unique identifiers).

organizationEventFilter_entityValues :: Lens' OrganizationEventFilter (Maybe (NonEmpty Text))

A list of entity identifiers, such as EC2 instance IDs (i-34ab692e) or EBS volumes (vol-426ab23e).

organizationEventFilter_services :: Lens' OrganizationEventFilter (Maybe (NonEmpty Text))

The AWS services associated with the event. For example, EC2, RDS.