Network.AWS.AWSHealth.Types

Description

Service Configuration

defaultService :: Service

API version 2016-08-04 of the Amazon Health APIs and Notifications SDK configuration.

Errors

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

EnableHealthServiceAccessForOrganization is already in progress. Wait for the action to complete before trying again. To get the current status, use the DescribeHealthServiceStatusForOrganization operation.

_InvalidPaginationToken :: AsError a => Getting (First ServiceError) a ServiceError

The specified pagination token (nextToken) is not valid.

_UnsupportedLocale :: AsError a => Getting (First ServiceError) a ServiceError

The specified locale is not supported.

EntityStatusCode

newtype EntityStatusCode

Constructors

EntityStatusCode'

Fields

fromEntityStatusCode :: Text

Bundled Patterns

pattern EntityStatusCode_IMPAIRED :: EntityStatusCode

pattern EntityStatusCode_UNIMPAIRED :: EntityStatusCode

pattern EntityStatusCode_UNKNOWN :: EntityStatusCode

EventAggregateField

newtype EventAggregateField

Constructors

EventAggregateField'

Fields

fromEventAggregateField :: Text

Bundled Patterns

pattern EventAggregateField_EventTypeCategory :: EventAggregateField

EventScopeCode

newtype EventScopeCode

Constructors

EventScopeCode'

Fields

fromEventScopeCode :: Text

Bundled Patterns

pattern EventScopeCode_ACCOUNT_SPECIFIC :: EventScopeCode

pattern EventScopeCode_NONE :: EventScopeCode

pattern EventScopeCode_PUBLIC :: EventScopeCode

EventStatusCode

newtype EventStatusCode

Constructors

EventStatusCode'

Fields

fromEventStatusCode :: Text

Bundled Patterns

pattern EventStatusCode_Closed :: EventStatusCode

pattern EventStatusCode_Open :: EventStatusCode

pattern EventStatusCode_Upcoming :: EventStatusCode

EventTypeCategory

newtype EventTypeCategory

Constructors

EventTypeCategory'

Fields

fromEventTypeCategory :: Text

Bundled Patterns

pattern EventTypeCategory_AccountNotification :: EventTypeCategory

pattern EventTypeCategory_Investigation :: EventTypeCategory

pattern EventTypeCategory_Issue :: EventTypeCategory

pattern EventTypeCategory_ScheduledChange :: EventTypeCategory

AffectedEntity

data AffectedEntity

Information about an entity that is affected by a Health event.

See: newAffectedEntity smart constructor.

Constructors

AffectedEntity'

Fields

awsAccountId :: Maybe Text

The 12-digit AWS account number that contains the affected entity.

eventArn :: Maybe Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

statusCode :: Maybe EntityStatusCode

The most recent status of the entity affected by the event. The possible values are IMPAIRED, UNIMPAIRED, and UNKNOWN.

tags :: Maybe (HashMap Text Text)

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

entityArn :: Maybe Text

The unique identifier for the entity. Format: arn:aws:health:entity-region:aws-account:entity/entity-id . Example: arn:aws:health:us-east-1:111222333444:entity/AVh5GGT7ul1arKr1sE1K

entityUrl :: Maybe Text

The URL of the affected entity.

entityValue :: Maybe Text

The ID of the affected entity.

lastUpdatedTime :: Maybe POSIX

The most recent time that the entity was updated.

newAffectedEntity :: AffectedEntity

Create a value of AffectedEntity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:AffectedEntity', affectedEntity_awsAccountId - The 12-digit AWS account number that contains the affected entity.

$sel:eventArn:AffectedEntity', affectedEntity_eventArn - The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

$sel:statusCode:AffectedEntity', affectedEntity_statusCode - The most recent status of the entity affected by the event. The possible values are IMPAIRED, UNIMPAIRED, and UNKNOWN.

$sel:tags:AffectedEntity', affectedEntity_tags - A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

$sel:entityArn:AffectedEntity', affectedEntity_entityArn - The unique identifier for the entity. Format: arn:aws:health:entity-region:aws-account:entity/entity-id . Example: arn:aws:health:us-east-1:111222333444:entity/AVh5GGT7ul1arKr1sE1K

$sel:entityUrl:AffectedEntity', affectedEntity_entityUrl - The URL of the affected entity.

$sel:entityValue:AffectedEntity', affectedEntity_entityValue - The ID of the affected entity.

$sel:lastUpdatedTime:AffectedEntity', affectedEntity_lastUpdatedTime - The most recent time that the entity was updated.

affectedEntity_awsAccountId :: Lens' AffectedEntity (Maybe Text)

The 12-digit AWS account number that contains the affected entity.

affectedEntity_eventArn :: Lens' AffectedEntity (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

affectedEntity_statusCode :: Lens' AffectedEntity (Maybe EntityStatusCode)

The most recent status of the entity affected by the event. The possible values are IMPAIRED, UNIMPAIRED, and UNKNOWN.

affectedEntity_tags :: Lens' AffectedEntity (Maybe (HashMap Text Text))

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

affectedEntity_entityArn :: Lens' AffectedEntity (Maybe Text)

The unique identifier for the entity. Format: arn:aws:health:entity-region:aws-account:entity/entity-id . Example: arn:aws:health:us-east-1:111222333444:entity/AVh5GGT7ul1arKr1sE1K

affectedEntity_entityUrl :: Lens' AffectedEntity (Maybe Text)

The URL of the affected entity.

affectedEntity_entityValue :: Lens' AffectedEntity (Maybe Text)

The ID of the affected entity.

affectedEntity_lastUpdatedTime :: Lens' AffectedEntity (Maybe UTCTime)

The most recent time that the entity was updated.

DateTimeRange

data DateTimeRange

A range of dates and times that is used by the EventFilter and EntityFilter objects. If from is set and to is set: match items where the timestamp (startTime, endTime, or lastUpdatedTime) is between from and to inclusive. If from is set and to is not set: match items where the timestamp value is equal to or after from. If from is not set and to is set: match items where the timestamp value is equal to or before to.

See: newDateTimeRange smart constructor.

Constructors

DateTimeRange'

Fields

to :: Maybe POSIX

The ending date and time of a time range.

from :: Maybe POSIX

The starting date and time of a time range.

newDateTimeRange :: DateTimeRange

Create a value of DateTimeRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:to:DateTimeRange', dateTimeRange_to - The ending date and time of a time range.

$sel:from:DateTimeRange', dateTimeRange_from - The starting date and time of a time range.

dateTimeRange_to :: Lens' DateTimeRange (Maybe UTCTime)

The ending date and time of a time range.

dateTimeRange_from :: Lens' DateTimeRange (Maybe UTCTime)

The starting date and time of a time range.

EntityAggregate

data EntityAggregate

The number of entities that are affected by one or more events. Returned by the DescribeEntityAggregates operation.

See: newEntityAggregate smart constructor.

Constructors

EntityAggregate'

Fields

eventArn :: Maybe Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

count :: Maybe Int

The number of entities that match the criteria for the specified events.

newEntityAggregate :: EntityAggregate

Create a value of EntityAggregate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventArn:EntityAggregate', entityAggregate_eventArn - The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

$sel:count:EntityAggregate', entityAggregate_count - The number of entities that match the criteria for the specified events.

entityAggregate_eventArn :: Lens' EntityAggregate (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

entityAggregate_count :: Lens' EntityAggregate (Maybe Int)

The number of entities that match the criteria for the specified events.

EntityFilter

data EntityFilter

The values to use to filter results from the EntityFilter operation.

See: newEntityFilter smart constructor.

Constructors

EntityFilter'

Fields

entityArns :: Maybe (NonEmpty Text)

A list of entity ARNs (unique identifiers).

statusCodes :: Maybe (NonEmpty EntityStatusCode)

A list of entity status codes (IMPAIRED, UNIMPAIRED, or UNKNOWN).

lastUpdatedTimes :: Maybe (NonEmpty DateTimeRange)

A list of the most recent dates and times that the entity was updated.

tags :: Maybe [HashMap Text Text]

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

entityValues :: Maybe (NonEmpty Text)

A list of IDs for affected entities.

eventArns :: NonEmpty Text

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

newEntityFilter

Arguments

:: NonEmpty Text
$sel:eventArns:EntityFilter'

-> EntityFilter

Create a value of EntityFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entityArns:EntityFilter', entityFilter_entityArns - A list of entity ARNs (unique identifiers).

$sel:statusCodes:EntityFilter', entityFilter_statusCodes - A list of entity status codes (IMPAIRED, UNIMPAIRED, or UNKNOWN).

$sel:lastUpdatedTimes:EntityFilter', entityFilter_lastUpdatedTimes - A list of the most recent dates and times that the entity was updated.

$sel:tags:EntityFilter', entityFilter_tags - A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

$sel:entityValues:EntityFilter', entityFilter_entityValues - A list of IDs for affected entities.

$sel:eventArns:EntityFilter', entityFilter_eventArns - A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

entityFilter_entityArns :: Lens' EntityFilter (Maybe (NonEmpty Text))

A list of entity ARNs (unique identifiers).

entityFilter_statusCodes :: Lens' EntityFilter (Maybe (NonEmpty EntityStatusCode))

A list of entity status codes (IMPAIRED, UNIMPAIRED, or UNKNOWN).

entityFilter_lastUpdatedTimes :: Lens' EntityFilter (Maybe (NonEmpty DateTimeRange))

A list of the most recent dates and times that the entity was updated.

entityFilter_tags :: Lens' EntityFilter (Maybe [HashMap Text Text])

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

entityFilter_entityValues :: Lens' EntityFilter (Maybe (NonEmpty Text))

A list of IDs for affected entities.

entityFilter_eventArns :: Lens' EntityFilter (NonEmpty Text)

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

Event

data Event

Summary information about an AWS Health event.

AWS Health events can be public or account-specific:

Public events might be service events that are not specific to an AWS account. For example, if there is an issue with an AWS Region, AWS Health provides information about the event, even if you don't use services or resources in that Region.

Account-specific events are specific to either your AWS account or an account in your organization. For example, if there's an issue with Amazon Elastic Compute Cloud in a Region that you use, AWS Health provides information about the event and the affected resources in the account.

You can determine if an event is public or account-specific by using the eventScopeCode parameter. For more information, see eventScopeCode.

See: newEvent smart constructor.

Constructors

Event'

Fields

eventTypeCategory :: Maybe EventTypeCategory

The category of the event. Possible values are issue, scheduledChange, and accountNotification.

eventScopeCode :: Maybe EventScopeCode

This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

startTime :: Maybe POSIX

The date and time that the event began.

arn :: Maybe Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

service :: Maybe Text

The AWS service that is affected by the event. For example, EC2, RDS.

endTime :: Maybe POSIX

The date and time that the event ended.

availabilityZone :: Maybe Text

The AWS Availability Zone of the event. For example, us-east-1a.

statusCode :: Maybe EventStatusCode

The most recent status of the event. Possible values are open, closed, and upcoming.

eventTypeCode :: Maybe Text

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

region :: Maybe Text

The AWS Region name of the event.

lastUpdatedTime :: Maybe POSIX

The most recent date and time that the event was updated.

newEvent :: Event

Create a value of Event with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventTypeCategory:Event', event_eventTypeCategory - The category of the event. Possible values are issue, scheduledChange, and accountNotification.

$sel:eventScopeCode:Event', event_eventScopeCode - This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

$sel:startTime:Event', event_startTime - The date and time that the event began.

$sel:arn:Event', event_arn - The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

$sel:service:Event', event_service - The AWS service that is affected by the event. For example, EC2, RDS.

$sel:endTime:Event', event_endTime - The date and time that the event ended.

$sel:availabilityZone:Event', event_availabilityZone - The AWS Availability Zone of the event. For example, us-east-1a.

$sel:statusCode:Event', event_statusCode - The most recent status of the event. Possible values are open, closed, and upcoming.

$sel:eventTypeCode:Event', event_eventTypeCode - The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

$sel:region:Event', event_region - The AWS Region name of the event.

$sel:lastUpdatedTime:Event', event_lastUpdatedTime - The most recent date and time that the event was updated.

event_eventTypeCategory :: Lens' Event (Maybe EventTypeCategory)

The category of the event. Possible values are issue, scheduledChange, and accountNotification.

event_eventScopeCode :: Lens' Event (Maybe EventScopeCode)

This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

event_startTime :: Lens' Event (Maybe UTCTime)

The date and time that the event began.

event_arn :: Lens' Event (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

event_service :: Lens' Event (Maybe Text)

The AWS service that is affected by the event. For example, EC2, RDS.

event_endTime :: Lens' Event (Maybe UTCTime)

The date and time that the event ended.

event_availabilityZone :: Lens' Event (Maybe Text)

The AWS Availability Zone of the event. For example, us-east-1a.

event_statusCode :: Lens' Event (Maybe EventStatusCode)

The most recent status of the event. Possible values are open, closed, and upcoming.

event_eventTypeCode :: Lens' Event (Maybe Text)

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

event_region :: Lens' Event (Maybe Text)

The AWS Region name of the event.

event_lastUpdatedTime :: Lens' Event (Maybe UTCTime)

The most recent date and time that the event was updated.

EventAccountFilter

data EventAccountFilter

The values used to filter results from the DescribeEventDetailsForOrganization and DescribeAffectedEntitiesForOrganization operations.

See: newEventAccountFilter smart constructor.

Constructors

EventAccountFilter'

Fields

awsAccountId :: Maybe Text

The 12-digit AWS account numbers that contains the affected entities.

eventArn :: Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

newEventAccountFilter

Arguments

:: Text
$sel:eventArn:EventAccountFilter'

-> EventAccountFilter

Create a value of EventAccountFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:EventAccountFilter', eventAccountFilter_awsAccountId - The 12-digit AWS account numbers that contains the affected entities.

$sel:eventArn:EventAccountFilter', eventAccountFilter_eventArn - The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

eventAccountFilter_awsAccountId :: Lens' EventAccountFilter (Maybe Text)

The 12-digit AWS account numbers that contains the affected entities.

eventAccountFilter_eventArn :: Lens' EventAccountFilter Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

EventAggregate

data EventAggregate

The number of events of each issue type. Returned by the DescribeEventAggregates operation.

See: newEventAggregate smart constructor.

Constructors

EventAggregate'

Fields

count :: Maybe Int

The number of events of the associated issue type.

aggregateValue :: Maybe Text

The issue type for the associated count.

newEventAggregate :: EventAggregate

Create a value of EventAggregate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:count:EventAggregate', eventAggregate_count - The number of events of the associated issue type.

$sel:aggregateValue:EventAggregate', eventAggregate_aggregateValue - The issue type for the associated count.

eventAggregate_count :: Lens' EventAggregate (Maybe Int)

The number of events of the associated issue type.

eventAggregate_aggregateValue :: Lens' EventAggregate (Maybe Text)

The issue type for the associated count.

EventDescription

data EventDescription

The detailed description of the event. Included in the information returned by the DescribeEventDetails operation.

See: newEventDescription smart constructor.

Constructors

EventDescription'

Fields

latestDescription :: Maybe Text

The most recent description of the event.

newEventDescription :: EventDescription

Create a value of EventDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestDescription:EventDescription', eventDescription_latestDescription - The most recent description of the event.

eventDescription_latestDescription :: Lens' EventDescription (Maybe Text)

The most recent description of the event.

EventDetails

data EventDetails

Detailed information about an event. A combination of an Event object, an EventDescription object, and additional metadata about the event. Returned by the DescribeEventDetails operation.

See: newEventDetails smart constructor.

Constructors

EventDetails'

Fields

eventMetadata :: Maybe (HashMap Text Text)

Additional metadata about the event.

eventDescription :: Maybe EventDescription

The most recent description of the event.

event :: Maybe Event

Summary information about the event.

newEventDetails :: EventDetails

Create a value of EventDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventMetadata:EventDetails', eventDetails_eventMetadata - Additional metadata about the event.

$sel:eventDescription:EventDetails', eventDetails_eventDescription - The most recent description of the event.

$sel:event:EventDetails', eventDetails_event - Summary information about the event.

eventDetails_eventMetadata :: Lens' EventDetails (Maybe (HashMap Text Text))

Additional metadata about the event.

eventDetails_eventDescription :: Lens' EventDetails (Maybe EventDescription)

The most recent description of the event.

eventDetails_event :: Lens' EventDetails (Maybe Event)

Summary information about the event.

EventDetailsErrorItem

data EventDetailsErrorItem

Error information returned when a DescribeEventDetails operation can't find a specified event.

See: newEventDetailsErrorItem smart constructor.

Constructors

EventDetailsErrorItem'

Fields

errorName :: Maybe Text

The name of the error.

eventArn :: Maybe Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

errorMessage :: Maybe Text

A message that describes the error.

newEventDetailsErrorItem :: EventDetailsErrorItem

Create a value of EventDetailsErrorItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorName:EventDetailsErrorItem', eventDetailsErrorItem_errorName - The name of the error.

$sel:eventArn:EventDetailsErrorItem', eventDetailsErrorItem_eventArn - The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

$sel:errorMessage:EventDetailsErrorItem', eventDetailsErrorItem_errorMessage - A message that describes the error.

eventDetailsErrorItem_errorName :: Lens' EventDetailsErrorItem (Maybe Text)

The name of the error.

eventDetailsErrorItem_eventArn :: Lens' EventDetailsErrorItem (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

eventDetailsErrorItem_errorMessage :: Lens' EventDetailsErrorItem (Maybe Text)

A message that describes the error.

EventFilter

data EventFilter

The values to use to filter results from the DescribeEvents and DescribeEventAggregates operations.

See: newEventFilter smart constructor.

Constructors

EventFilter'

Fields

availabilityZones :: Maybe [Text]

A list of AWS Availability Zones.

endTimes :: Maybe (NonEmpty DateTimeRange)

A list of dates and times that the event ended.

startTimes :: Maybe (NonEmpty DateTimeRange)

A list of dates and times that the event began.

services :: Maybe (NonEmpty Text)

The AWS services associated with the event. For example, EC2, RDS.

entityArns :: Maybe (NonEmpty Text)

A list of entity ARNs (unique identifiers).

eventTypeCodes :: Maybe (NonEmpty Text)

A list of unique identifiers for event types. For example, "AWS_EC2_SYSTEM_MAINTENANCE_EVENT","AWS_RDS_MAINTENANCE_SCHEDULED".

lastUpdatedTimes :: Maybe (NonEmpty DateTimeRange)

A list of dates and times that the event was last updated.

tags :: Maybe [HashMap Text Text]

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

eventStatusCodes :: Maybe (NonEmpty EventStatusCode)

A list of event status codes.

regions :: Maybe (NonEmpty Text)

A list of AWS Regions.

entityValues :: Maybe (NonEmpty Text)

A list of entity identifiers, such as EC2 instance IDs (i-34ab692e) or EBS volumes (vol-426ab23e).

eventTypeCategories :: Maybe (NonEmpty EventTypeCategory)

A list of event type category codes (issue, scheduledChange, or accountNotification).

eventArns :: Maybe (NonEmpty Text)

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

newEventFilter :: EventFilter

Create a value of EventFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZones:EventFilter', eventFilter_availabilityZones - A list of AWS Availability Zones.

$sel:endTimes:EventFilter', eventFilter_endTimes - A list of dates and times that the event ended.

$sel:startTimes:EventFilter', eventFilter_startTimes - A list of dates and times that the event began.

$sel:services:EventFilter', eventFilter_services - The AWS services associated with the event. For example, EC2, RDS.

$sel:entityArns:EventFilter', eventFilter_entityArns - A list of entity ARNs (unique identifiers).

$sel:eventTypeCodes:EventFilter', eventFilter_eventTypeCodes - A list of unique identifiers for event types. For example, "AWS_EC2_SYSTEM_MAINTENANCE_EVENT","AWS_RDS_MAINTENANCE_SCHEDULED".

$sel:lastUpdatedTimes:EventFilter', eventFilter_lastUpdatedTimes - A list of dates and times that the event was last updated.

$sel:tags:EventFilter', eventFilter_tags - A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

$sel:eventStatusCodes:EventFilter', eventFilter_eventStatusCodes - A list of event status codes.

$sel:regions:EventFilter', eventFilter_regions - A list of AWS Regions.

$sel:entityValues:EventFilter', eventFilter_entityValues - A list of entity identifiers, such as EC2 instance IDs (i-34ab692e) or EBS volumes (vol-426ab23e).

$sel:eventTypeCategories:EventFilter', eventFilter_eventTypeCategories - A list of event type category codes (issue, scheduledChange, or accountNotification).

$sel:eventArns:EventFilter', eventFilter_eventArns - A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

eventFilter_availabilityZones :: Lens' EventFilter (Maybe [Text])

A list of AWS Availability Zones.

eventFilter_endTimes :: Lens' EventFilter (Maybe (NonEmpty DateTimeRange))

A list of dates and times that the event ended.

eventFilter_startTimes :: Lens' EventFilter (Maybe (NonEmpty DateTimeRange))

A list of dates and times that the event began.

eventFilter_services :: Lens' EventFilter (Maybe (NonEmpty Text))

The AWS services associated with the event. For example, EC2, RDS.

eventFilter_entityArns :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of entity ARNs (unique identifiers).

eventFilter_eventTypeCodes :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of unique identifiers for event types. For example, "AWS_EC2_SYSTEM_MAINTENANCE_EVENT","AWS_RDS_MAINTENANCE_SCHEDULED".

eventFilter_lastUpdatedTimes :: Lens' EventFilter (Maybe (NonEmpty DateTimeRange))

A list of dates and times that the event was last updated.

eventFilter_tags :: Lens' EventFilter (Maybe [HashMap Text Text])

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

eventFilter_eventStatusCodes :: Lens' EventFilter (Maybe (NonEmpty EventStatusCode))

A list of event status codes.

eventFilter_regions :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of AWS Regions.

eventFilter_entityValues :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of entity identifiers, such as EC2 instance IDs (i-34ab692e) or EBS volumes (vol-426ab23e).

eventFilter_eventTypeCategories :: Lens' EventFilter (Maybe (NonEmpty EventTypeCategory))

A list of event type category codes (issue, scheduledChange, or accountNotification).

eventFilter_eventArns :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

EventType

data EventType

Contains the metadata about a type of event that is reported by AWS Health. The EventType shows the category, service, and the event type code of the event. For example, an issue might be the category, EC2 the service, and AWS_EC2_SYSTEM_MAINTENANCE_EVENT the event type code.

You can use the DescribeEventTypes API operation to return this information about an event.

You can also use the Amazon CloudWatch Events console to create a rule so that you can get notified or take action when AWS Health delivers a specific event to your AWS account. For more information, see Monitor for AWS Health events with Amazon CloudWatch Events in the AWS Health User Guide.

See: newEventType smart constructor.

Constructors

EventType'

Fields

category :: Maybe EventTypeCategory

A list of event type category codes (issue, scheduledChange, or accountNotification).

code :: Maybe Text

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

service :: Maybe Text

The AWS service that is affected by the event. For example, EC2, RDS.

newEventType :: EventType

Create a value of EventType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:category:EventType', eventType_category - A list of event type category codes (issue, scheduledChange, or accountNotification).

$sel:code:EventType', eventType_code - The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

$sel:service:EventType', eventType_service - The AWS service that is affected by the event. For example, EC2, RDS.

eventType_category :: Lens' EventType (Maybe EventTypeCategory)

A list of event type category codes (issue, scheduledChange, or accountNotification).

eventType_code :: Lens' EventType (Maybe Text)

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION ; for example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

eventType_service :: Lens' EventType (Maybe Text)

The AWS service that is affected by the event. For example, EC2, RDS.

EventTypeFilter

data EventTypeFilter

The values to use to filter results from the DescribeEventTypes operation.

See: newEventTypeFilter smart constructor.

Constructors

EventTypeFilter'

Fields

services :: Maybe (NonEmpty Text)

The AWS services associated with the event. For example, EC2, RDS.

eventTypeCodes :: Maybe (NonEmpty Text)

A list of event type codes.

eventTypeCategories :: Maybe (NonEmpty EventTypeCategory)

A list of event type category codes (issue, scheduledChange, or accountNotification).

newEventTypeFilter :: EventTypeFilter

Create a value of EventTypeFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:services:EventTypeFilter', eventTypeFilter_services - The AWS services associated with the event. For example, EC2, RDS.

$sel:eventTypeCodes:EventTypeFilter', eventTypeFilter_eventTypeCodes - A list of event type codes.

$sel:eventTypeCategories:EventTypeFilter', eventTypeFilter_eventTypeCategories - A list of event type category codes (issue, scheduledChange, or accountNotification).

eventTypeFilter_services :: Lens' EventTypeFilter (Maybe (NonEmpty Text))

The AWS services associated with the event. For example, EC2, RDS.

eventTypeFilter_eventTypeCodes :: Lens' EventTypeFilter (Maybe (NonEmpty Text))

A list of event type codes.

eventTypeFilter_eventTypeCategories :: Lens' EventTypeFilter (Maybe (NonEmpty EventTypeCategory))

A list of event type category codes (issue, scheduledChange, or accountNotification).

OrganizationAffectedEntitiesErrorItem

data OrganizationAffectedEntitiesErrorItem

Error information returned when a DescribeAffectedEntitiesForOrganization operation can't find or process a specific entity.

See: newOrganizationAffectedEntitiesErrorItem smart constructor.

Constructors

OrganizationAffectedEntitiesErrorItem'

Fields

errorName :: Maybe Text

The name of the error.

awsAccountId :: Maybe Text

The 12-digit AWS account numbers that contains the affected entities.

eventArn :: Maybe Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

errorMessage :: Maybe Text

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

newOrganizationAffectedEntitiesErrorItem :: OrganizationAffectedEntitiesErrorItem

Create a value of OrganizationAffectedEntitiesErrorItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorName:OrganizationAffectedEntitiesErrorItem', organizationAffectedEntitiesErrorItem_errorName - The name of the error.

$sel:awsAccountId:OrganizationAffectedEntitiesErrorItem', organizationAffectedEntitiesErrorItem_awsAccountId - The 12-digit AWS account numbers that contains the affected entities.

$sel:eventArn:OrganizationAffectedEntitiesErrorItem', organizationAffectedEntitiesErrorItem_eventArn - The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

$sel:errorMessage:OrganizationAffectedEntitiesErrorItem', organizationAffectedEntitiesErrorItem_errorMessage - The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

organizationAffectedEntitiesErrorItem_errorName :: Lens' OrganizationAffectedEntitiesErrorItem (Maybe Text)

The name of the error.

organizationAffectedEntitiesErrorItem_awsAccountId :: Lens' OrganizationAffectedEntitiesErrorItem (Maybe Text)

The 12-digit AWS account numbers that contains the affected entities.

organizationAffectedEntitiesErrorItem_eventArn :: Lens' OrganizationAffectedEntitiesErrorItem (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

organizationAffectedEntitiesErrorItem_errorMessage :: Lens' OrganizationAffectedEntitiesErrorItem (Maybe Text)

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

OrganizationEvent

data OrganizationEvent

Summary information about an event, returned by the DescribeEventsForOrganization operation.

See: newOrganizationEvent smart constructor.

Constructors

OrganizationEvent'

Fields

eventTypeCategory :: Maybe EventTypeCategory

The category of the event type.

eventScopeCode :: Maybe EventScopeCode

This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

startTime :: Maybe POSIX

The date and time that the event began.

arn :: Maybe Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

service :: Maybe Text

The AWS service that is affected by the event, such as EC2 and RDS.

endTime :: Maybe POSIX

The date and time that the event ended.

statusCode :: Maybe EventStatusCode

The most recent status of the event. Possible values are open, closed, and upcoming.

eventTypeCode :: Maybe Text

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

region :: Maybe Text

The AWS Region name of the event.

lastUpdatedTime :: Maybe POSIX

The most recent date and time that the event was updated.

newOrganizationEvent :: OrganizationEvent

Create a value of OrganizationEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventTypeCategory:OrganizationEvent', organizationEvent_eventTypeCategory - The category of the event type.

$sel:eventScopeCode:OrganizationEvent', organizationEvent_eventScopeCode - This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

$sel:startTime:OrganizationEvent', organizationEvent_startTime - The date and time that the event began.

$sel:arn:OrganizationEvent', organizationEvent_arn - The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

$sel:service:OrganizationEvent', organizationEvent_service - The AWS service that is affected by the event, such as EC2 and RDS.

$sel:endTime:OrganizationEvent', organizationEvent_endTime - The date and time that the event ended.

$sel:statusCode:OrganizationEvent', organizationEvent_statusCode - The most recent status of the event. Possible values are open, closed, and upcoming.

$sel:eventTypeCode:OrganizationEvent', organizationEvent_eventTypeCode - The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

$sel:region:OrganizationEvent', organizationEvent_region - The AWS Region name of the event.

$sel:lastUpdatedTime:OrganizationEvent', organizationEvent_lastUpdatedTime - The most recent date and time that the event was updated.

organizationEvent_eventTypeCategory :: Lens' OrganizationEvent (Maybe EventTypeCategory)

The category of the event type.

organizationEvent_eventScopeCode :: Lens' OrganizationEvent (Maybe EventScopeCode)

This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

organizationEvent_startTime :: Lens' OrganizationEvent (Maybe UTCTime)

The date and time that the event began.

organizationEvent_arn :: Lens' OrganizationEvent (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

organizationEvent_service :: Lens' OrganizationEvent (Maybe Text)

The AWS service that is affected by the event, such as EC2 and RDS.

organizationEvent_endTime :: Lens' OrganizationEvent (Maybe UTCTime)

The date and time that the event ended.

organizationEvent_statusCode :: Lens' OrganizationEvent (Maybe EventStatusCode)

The most recent status of the event. Possible values are open, closed, and upcoming.

organizationEvent_eventTypeCode :: Lens' OrganizationEvent (Maybe Text)

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

organizationEvent_region :: Lens' OrganizationEvent (Maybe Text)

The AWS Region name of the event.

organizationEvent_lastUpdatedTime :: Lens' OrganizationEvent (Maybe UTCTime)

The most recent date and time that the event was updated.

OrganizationEventDetails

data OrganizationEventDetails

Detailed information about an event. A combination of an Event object, an EventDescription object, and additional metadata about the event. Returned by the DescribeEventDetailsForOrganization operation.

See: newOrganizationEventDetails smart constructor.

Constructors

OrganizationEventDetails'

Fields

awsAccountId :: Maybe Text

The 12-digit AWS account numbers that contains the affected entities.

eventMetadata :: Maybe (HashMap Text Text)

Additional metadata about the event.

eventDescription :: Maybe EventDescription

event :: Maybe Event

newOrganizationEventDetails :: OrganizationEventDetails

Create a value of OrganizationEventDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:OrganizationEventDetails', organizationEventDetails_awsAccountId - The 12-digit AWS account numbers that contains the affected entities.

$sel:eventMetadata:OrganizationEventDetails', organizationEventDetails_eventMetadata - Additional metadata about the event.

$sel:eventDescription:OrganizationEventDetails', organizationEventDetails_eventDescription - Undocumented member.

$sel:event:OrganizationEventDetails', organizationEventDetails_event - Undocumented member.

organizationEventDetails_awsAccountId :: Lens' OrganizationEventDetails (Maybe Text)

The 12-digit AWS account numbers that contains the affected entities.

organizationEventDetails_eventMetadata :: Lens' OrganizationEventDetails (Maybe (HashMap Text Text))

Additional metadata about the event.

organizationEventDetails_eventDescription :: Lens' OrganizationEventDetails (Maybe EventDescription)

Undocumented member.

organizationEventDetails_event :: Lens' OrganizationEventDetails (Maybe Event)

Undocumented member.

OrganizationEventDetailsErrorItem

data OrganizationEventDetailsErrorItem

Error information returned when a DescribeEventDetailsForOrganization operation can't find a specified event.

See: newOrganizationEventDetailsErrorItem smart constructor.

Constructors

OrganizationEventDetailsErrorItem'

Fields

errorName :: Maybe Text

The name of the error.

awsAccountId :: Maybe Text

Error information returned when a DescribeEventDetailsForOrganization operation can't find a specified event.

eventArn :: Maybe Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

errorMessage :: Maybe Text

A message that describes the error.

If you call the DescribeEventDetailsForOrganization operation and receive one of the following errors, follow the recommendations in the message:

We couldn't find a public event that matches your request. To find an event that is account specific, you must enter an AWS account ID in the request.

We couldn't find an account specific event for the specified AWS account. To find an event that is public, you must enter a null value for the AWS account ID in the request.

Your AWS account doesn't include the AWS Support plan required to use the AWS Health API. You must have either a Business or Enterprise Support plan.

newOrganizationEventDetailsErrorItem :: OrganizationEventDetailsErrorItem

Create a value of OrganizationEventDetailsErrorItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorName:OrganizationEventDetailsErrorItem', organizationEventDetailsErrorItem_errorName - The name of the error.

$sel:awsAccountId:OrganizationEventDetailsErrorItem', organizationEventDetailsErrorItem_awsAccountId - Error information returned when a DescribeEventDetailsForOrganization operation can't find a specified event.

$sel:eventArn:OrganizationEventDetailsErrorItem', organizationEventDetailsErrorItem_eventArn - The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

$sel:errorMessage:OrganizationEventDetailsErrorItem', organizationEventDetailsErrorItem_errorMessage - A message that describes the error.

If you call the DescribeEventDetailsForOrganization operation and receive one of the following errors, follow the recommendations in the message:

We couldn't find a public event that matches your request. To find an event that is account specific, you must enter an AWS account ID in the request.

We couldn't find an account specific event for the specified AWS account. To find an event that is public, you must enter a null value for the AWS account ID in the request.

Your AWS account doesn't include the AWS Support plan required to use the AWS Health API. You must have either a Business or Enterprise Support plan.

organizationEventDetailsErrorItem_errorName :: Lens' OrganizationEventDetailsErrorItem (Maybe Text)

The name of the error.

organizationEventDetailsErrorItem_awsAccountId :: Lens' OrganizationEventDetailsErrorItem (Maybe Text)

Error information returned when a DescribeEventDetailsForOrganization operation can't find a specified event.

organizationEventDetailsErrorItem_eventArn :: Lens' OrganizationEventDetailsErrorItem (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

organizationEventDetailsErrorItem_errorMessage :: Lens' OrganizationEventDetailsErrorItem (Maybe Text)

A message that describes the error.

If you call the DescribeEventDetailsForOrganization operation and receive one of the following errors, follow the recommendations in the message:

We couldn't find a public event that matches your request. To find an event that is account specific, you must enter an AWS account ID in the request.

We couldn't find an account specific event for the specified AWS account. To find an event that is public, you must enter a null value for the AWS account ID in the request.

Your AWS account doesn't include the AWS Support plan required to use the AWS Health API. You must have either a Business or Enterprise Support plan.

OrganizationEventFilter

data OrganizationEventFilter

The values to filter results from the DescribeEventsForOrganization operation.

See: newOrganizationEventFilter smart constructor.

Constructors

OrganizationEventFilter'

Fields

services :: Maybe (NonEmpty Text)

The AWS services associated with the event. For example, EC2, RDS.

startTime :: Maybe DateTimeRange

entityArns :: Maybe (NonEmpty Text)

A list of entity ARNs (unique identifiers).

eventTypeCodes :: Maybe (NonEmpty Text)

A list of unique identifiers for event types. For example, "AWS_EC2_SYSTEM_MAINTENANCE_EVENT","AWS_RDS_MAINTENANCE_SCHEDULED".

endTime :: Maybe DateTimeRange

eventStatusCodes :: Maybe (NonEmpty EventStatusCode)

A list of event status codes.

regions :: Maybe (NonEmpty Text)

A list of AWS Regions.

entityValues :: Maybe (NonEmpty Text)

A list of entity identifiers, such as EC2 instance IDs (i-34ab692e) or EBS volumes (vol-426ab23e).

eventTypeCategories :: Maybe (NonEmpty EventTypeCategory)

A list of event type category codes (issue, scheduledChange, or accountNotification).

lastUpdatedTime :: Maybe DateTimeRange

awsAccountIds :: Maybe (NonEmpty Text)

A list of 12-digit AWS account numbers that contains the affected entities.

newOrganizationEventFilter :: OrganizationEventFilter

Create a value of OrganizationEventFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:services:OrganizationEventFilter', organizationEventFilter_services - The AWS services associated with the event. For example, EC2, RDS.

$sel:startTime:OrganizationEventFilter', organizationEventFilter_startTime - Undocumented member.

$sel:entityArns:OrganizationEventFilter', organizationEventFilter_entityArns - A list of entity ARNs (unique identifiers).

$sel:eventTypeCodes:OrganizationEventFilter', organizationEventFilter_eventTypeCodes - A list of unique identifiers for event types. For example, "AWS_EC2_SYSTEM_MAINTENANCE_EVENT","AWS_RDS_MAINTENANCE_SCHEDULED".

$sel:endTime:OrganizationEventFilter', organizationEventFilter_endTime - Undocumented member.

$sel:eventStatusCodes:OrganizationEventFilter', organizationEventFilter_eventStatusCodes - A list of event status codes.

$sel:regions:OrganizationEventFilter', organizationEventFilter_regions - A list of AWS Regions.

$sel:entityValues:OrganizationEventFilter', organizationEventFilter_entityValues - A list of entity identifiers, such as EC2 instance IDs (i-34ab692e) or EBS volumes (vol-426ab23e).

$sel:eventTypeCategories:OrganizationEventFilter', organizationEventFilter_eventTypeCategories - A list of event type category codes (issue, scheduledChange, or accountNotification).

$sel:lastUpdatedTime:OrganizationEventFilter', organizationEventFilter_lastUpdatedTime - Undocumented member.

$sel:awsAccountIds:OrganizationEventFilter', organizationEventFilter_awsAccountIds - A list of 12-digit AWS account numbers that contains the affected entities.

organizationEventFilter_services :: Lens' OrganizationEventFilter (Maybe (NonEmpty Text))

The AWS services associated with the event. For example, EC2, RDS.

organizationEventFilter_startTime :: Lens' OrganizationEventFilter (Maybe DateTimeRange)

Undocumented member.

organizationEventFilter_entityArns :: Lens' OrganizationEventFilter (Maybe (NonEmpty Text))

A list of entity ARNs (unique identifiers).

organizationEventFilter_eventTypeCodes :: Lens' OrganizationEventFilter (Maybe (NonEmpty Text))

A list of unique identifiers for event types. For example, "AWS_EC2_SYSTEM_MAINTENANCE_EVENT","AWS_RDS_MAINTENANCE_SCHEDULED".

organizationEventFilter_endTime :: Lens' OrganizationEventFilter (Maybe DateTimeRange)

Undocumented member.

organizationEventFilter_eventStatusCodes :: Lens' OrganizationEventFilter (Maybe (NonEmpty EventStatusCode))

A list of event status codes.

organizationEventFilter_regions :: Lens' OrganizationEventFilter (Maybe (NonEmpty Text))

A list of AWS Regions.

organizationEventFilter_entityValues :: Lens' OrganizationEventFilter (Maybe (NonEmpty Text))

A list of entity identifiers, such as EC2 instance IDs (i-34ab692e) or EBS volumes (vol-426ab23e).

organizationEventFilter_eventTypeCategories :: Lens' OrganizationEventFilter (Maybe (NonEmpty EventTypeCategory))

A list of event type category codes (issue, scheduledChange, or accountNotification).

organizationEventFilter_lastUpdatedTime :: Lens' OrganizationEventFilter (Maybe DateTimeRange)

Undocumented member.

organizationEventFilter_awsAccountIds :: Lens' OrganizationEventFilter (Maybe (NonEmpty Text))

A list of 12-digit AWS account numbers that contains the affected entities.