Network.AWS.AWSHealth.Types.AffectedEntity

Description

Documentation

data AffectedEntity

Information about an entity that is affected by a Health event.

See: newAffectedEntity smart constructor.

Constructors

AffectedEntity'

Fields

awsAccountId :: Maybe Text

The 12-digit AWS account number that contains the affected entity.

eventArn :: Maybe Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

statusCode :: Maybe EntityStatusCode

The most recent status of the entity affected by the event. The possible values are IMPAIRED, UNIMPAIRED, and UNKNOWN.

tags :: Maybe (HashMap Text Text)

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

entityArn :: Maybe Text

The unique identifier for the entity. Format: arn:aws:health:entity-region:aws-account:entity/entity-id . Example: arn:aws:health:us-east-1:111222333444:entity/AVh5GGT7ul1arKr1sE1K

entityUrl :: Maybe Text

The URL of the affected entity.

entityValue :: Maybe Text

The ID of the affected entity.

lastUpdatedTime :: Maybe POSIX

The most recent time that the entity was updated.

newAffectedEntity :: AffectedEntity

Create a value of AffectedEntity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:AffectedEntity', affectedEntity_awsAccountId - The 12-digit AWS account number that contains the affected entity.

$sel:eventArn:AffectedEntity', affectedEntity_eventArn - The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

$sel:statusCode:AffectedEntity', affectedEntity_statusCode - The most recent status of the entity affected by the event. The possible values are IMPAIRED, UNIMPAIRED, and UNKNOWN.

$sel:tags:AffectedEntity', affectedEntity_tags - A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

$sel:entityArn:AffectedEntity', affectedEntity_entityArn - The unique identifier for the entity. Format: arn:aws:health:entity-region:aws-account:entity/entity-id . Example: arn:aws:health:us-east-1:111222333444:entity/AVh5GGT7ul1arKr1sE1K

$sel:entityUrl:AffectedEntity', affectedEntity_entityUrl - The URL of the affected entity.

$sel:entityValue:AffectedEntity', affectedEntity_entityValue - The ID of the affected entity.

$sel:lastUpdatedTime:AffectedEntity', affectedEntity_lastUpdatedTime - The most recent time that the entity was updated.

affectedEntity_awsAccountId :: Lens' AffectedEntity (Maybe Text)

The 12-digit AWS account number that contains the affected entity.

affectedEntity_eventArn :: Lens' AffectedEntity (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

affectedEntity_statusCode :: Lens' AffectedEntity (Maybe EntityStatusCode)

The most recent status of the entity affected by the event. The possible values are IMPAIRED, UNIMPAIRED, and UNKNOWN.

affectedEntity_tags :: Lens' AffectedEntity (Maybe (HashMap Text Text))

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

affectedEntity_entityArn :: Lens' AffectedEntity (Maybe Text)

The unique identifier for the entity. Format: arn:aws:health:entity-region:aws-account:entity/entity-id . Example: arn:aws:health:us-east-1:111222333444:entity/AVh5GGT7ul1arKr1sE1K

affectedEntity_entityUrl :: Lens' AffectedEntity (Maybe Text)

The URL of the affected entity.

affectedEntity_entityValue :: Lens' AffectedEntity (Maybe Text)

The ID of the affected entity.

affectedEntity_lastUpdatedTime :: Lens' AffectedEntity (Maybe UTCTime)

The most recent time that the entity was updated.