Network.AWS.AWSHealth.Types.EntityFilter

Description

Documentation

data EntityFilter

The values to use to filter results from the EntityFilter operation.

See: newEntityFilter smart constructor.

Constructors

EntityFilter'

Fields

entityArns :: Maybe (NonEmpty Text)

A list of entity ARNs (unique identifiers).

statusCodes :: Maybe (NonEmpty EntityStatusCode)

A list of entity status codes (IMPAIRED, UNIMPAIRED, or UNKNOWN).

lastUpdatedTimes :: Maybe (NonEmpty DateTimeRange)

A list of the most recent dates and times that the entity was updated.

tags :: Maybe [HashMap Text Text]

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

entityValues :: Maybe (NonEmpty Text)

A list of IDs for affected entities.

eventArns :: NonEmpty Text

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

newEntityFilter

Arguments

:: NonEmpty Text
$sel:eventArns:EntityFilter'

-> EntityFilter

Create a value of EntityFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entityArns:EntityFilter', entityFilter_entityArns - A list of entity ARNs (unique identifiers).

$sel:statusCodes:EntityFilter', entityFilter_statusCodes - A list of entity status codes (IMPAIRED, UNIMPAIRED, or UNKNOWN).

$sel:lastUpdatedTimes:EntityFilter', entityFilter_lastUpdatedTimes - A list of the most recent dates and times that the entity was updated.

$sel:tags:EntityFilter', entityFilter_tags - A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

$sel:entityValues:EntityFilter', entityFilter_entityValues - A list of IDs for affected entities.

$sel:eventArns:EntityFilter', entityFilter_eventArns - A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

entityFilter_entityArns :: Lens' EntityFilter (Maybe (NonEmpty Text))

A list of entity ARNs (unique identifiers).

entityFilter_statusCodes :: Lens' EntityFilter (Maybe (NonEmpty EntityStatusCode))

A list of entity status codes (IMPAIRED, UNIMPAIRED, or UNKNOWN).

entityFilter_lastUpdatedTimes :: Lens' EntityFilter (Maybe (NonEmpty DateTimeRange))

A list of the most recent dates and times that the entity was updated.

entityFilter_tags :: Lens' EntityFilter (Maybe [HashMap Text Text])

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

entityFilter_entityValues :: Lens' EntityFilter (Maybe (NonEmpty Text))

A list of IDs for affected entities.

entityFilter_eventArns :: Lens' EntityFilter (NonEmpty Text)

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"