Network.AWS.AWSHealth.Types.EventFilter

Description

Documentation

data EventFilter

The values to use to filter results from the DescribeEvents and DescribeEventAggregates operations.

See: newEventFilter smart constructor.

Constructors

EventFilter'

Fields

availabilityZones :: Maybe [Text]

A list of AWS Availability Zones.

endTimes :: Maybe (NonEmpty DateTimeRange)

A list of dates and times that the event ended.

startTimes :: Maybe (NonEmpty DateTimeRange)

A list of dates and times that the event began.

services :: Maybe (NonEmpty Text)

The AWS services associated with the event. For example, EC2, RDS.

entityArns :: Maybe (NonEmpty Text)

A list of entity ARNs (unique identifiers).

eventTypeCodes :: Maybe (NonEmpty Text)

A list of unique identifiers for event types. For example, "AWS_EC2_SYSTEM_MAINTENANCE_EVENT","AWS_RDS_MAINTENANCE_SCHEDULED".

lastUpdatedTimes :: Maybe (NonEmpty DateTimeRange)

A list of dates and times that the event was last updated.

tags :: Maybe [HashMap Text Text]

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

eventStatusCodes :: Maybe (NonEmpty EventStatusCode)

A list of event status codes.

regions :: Maybe (NonEmpty Text)

A list of AWS Regions.

entityValues :: Maybe (NonEmpty Text)

A list of entity identifiers, such as EC2 instance IDs (i-34ab692e) or EBS volumes (vol-426ab23e).

eventTypeCategories :: Maybe (NonEmpty EventTypeCategory)

A list of event type category codes (issue, scheduledChange, or accountNotification).

eventArns :: Maybe (NonEmpty Text)

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

newEventFilter :: EventFilter

Create a value of EventFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZones:EventFilter', eventFilter_availabilityZones - A list of AWS Availability Zones.

$sel:endTimes:EventFilter', eventFilter_endTimes - A list of dates and times that the event ended.

$sel:startTimes:EventFilter', eventFilter_startTimes - A list of dates and times that the event began.

$sel:services:EventFilter', eventFilter_services - The AWS services associated with the event. For example, EC2, RDS.

$sel:entityArns:EventFilter', eventFilter_entityArns - A list of entity ARNs (unique identifiers).

$sel:eventTypeCodes:EventFilter', eventFilter_eventTypeCodes - A list of unique identifiers for event types. For example, "AWS_EC2_SYSTEM_MAINTENANCE_EVENT","AWS_RDS_MAINTENANCE_SCHEDULED".

$sel:lastUpdatedTimes:EventFilter', eventFilter_lastUpdatedTimes - A list of dates and times that the event was last updated.

$sel:tags:EventFilter', eventFilter_tags - A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

$sel:eventStatusCodes:EventFilter', eventFilter_eventStatusCodes - A list of event status codes.

$sel:regions:EventFilter', eventFilter_regions - A list of AWS Regions.

$sel:entityValues:EventFilter', eventFilter_entityValues - A list of entity identifiers, such as EC2 instance IDs (i-34ab692e) or EBS volumes (vol-426ab23e).

$sel:eventTypeCategories:EventFilter', eventFilter_eventTypeCategories - A list of event type category codes (issue, scheduledChange, or accountNotification).

$sel:eventArns:EventFilter', eventFilter_eventArns - A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"

eventFilter_availabilityZones :: Lens' EventFilter (Maybe [Text])

A list of AWS Availability Zones.

eventFilter_endTimes :: Lens' EventFilter (Maybe (NonEmpty DateTimeRange))

A list of dates and times that the event ended.

eventFilter_startTimes :: Lens' EventFilter (Maybe (NonEmpty DateTimeRange))

A list of dates and times that the event began.

eventFilter_services :: Lens' EventFilter (Maybe (NonEmpty Text))

The AWS services associated with the event. For example, EC2, RDS.

eventFilter_entityArns :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of entity ARNs (unique identifiers).

eventFilter_eventTypeCodes :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of unique identifiers for event types. For example, "AWS_EC2_SYSTEM_MAINTENANCE_EVENT","AWS_RDS_MAINTENANCE_SCHEDULED".

eventFilter_lastUpdatedTimes :: Lens' EventFilter (Maybe (NonEmpty DateTimeRange))

A list of dates and times that the event was last updated.

eventFilter_tags :: Lens' EventFilter (Maybe [HashMap Text Text])

A map of entity tags attached to the affected entity.

Currently, the tags property isn't supported.

eventFilter_eventStatusCodes :: Lens' EventFilter (Maybe (NonEmpty EventStatusCode))

A list of event status codes.

eventFilter_regions :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of AWS Regions.

eventFilter_entityValues :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of entity identifiers, such as EC2 instance IDs (i-34ab692e) or EBS volumes (vol-426ab23e).

eventFilter_eventTypeCategories :: Lens' EventFilter (Maybe (NonEmpty EventTypeCategory))

A list of event type category codes (issue, scheduledChange, or accountNotification).

eventFilter_eventArns :: Lens' EventFilter (Maybe (NonEmpty Text))

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-CDE456", "arn:aws:health:us-west-1::event/EBS/AWS_EBS_LOST_VOLUME/AWS_EBS_LOST_VOLUME_CHI789_JKL101"