Network.AWS.AWSHealth.Types.OrganizationEvent

Description

Documentation

data OrganizationEvent

Summary information about an event, returned by the DescribeEventsForOrganization operation.

See: newOrganizationEvent smart constructor.

Constructors

OrganizationEvent'

Fields

lastUpdatedTime :: Maybe POSIX

The most recent date and time that the event was updated.

arn :: Maybe Text

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

service :: Maybe Text

The AWS service that is affected by the event, such as EC2 and RDS.

startTime :: Maybe POSIX

The date and time that the event began.

eventScopeCode :: Maybe EventScopeCode

This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

eventTypeCode :: Maybe Text

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

eventTypeCategory :: Maybe EventTypeCategory

The category of the event type.

endTime :: Maybe POSIX

The date and time that the event ended.

region :: Maybe Text

The AWS Region name of the event.

statusCode :: Maybe EventStatusCode

The most recent status of the event. Possible values are open, closed, and upcoming.

newOrganizationEvent :: OrganizationEvent

Create a value of OrganizationEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:OrganizationEvent', organizationEvent_lastUpdatedTime - The most recent date and time that the event was updated.

$sel:arn:OrganizationEvent', organizationEvent_arn - The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

$sel:service:OrganizationEvent', organizationEvent_service - The AWS service that is affected by the event, such as EC2 and RDS.

$sel:startTime:OrganizationEvent', organizationEvent_startTime - The date and time that the event began.

$sel:eventScopeCode:OrganizationEvent', organizationEvent_eventScopeCode - This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

$sel:eventTypeCode:OrganizationEvent', organizationEvent_eventTypeCode - The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

$sel:eventTypeCategory:OrganizationEvent', organizationEvent_eventTypeCategory - The category of the event type.

$sel:endTime:OrganizationEvent', organizationEvent_endTime - The date and time that the event ended.

$sel:region:OrganizationEvent', organizationEvent_region - The AWS Region name of the event.

$sel:statusCode:OrganizationEvent', organizationEvent_statusCode - The most recent status of the event. Possible values are open, closed, and upcoming.

organizationEvent_lastUpdatedTime :: Lens' OrganizationEvent (Maybe UTCTime)

The most recent date and time that the event was updated.

organizationEvent_arn :: Lens' OrganizationEvent (Maybe Text)

The unique identifier for the event. The event ARN has the arn:aws:health:event-region::event/SERVICE/EVENT_TYPE_CODE/EVENT_TYPE_PLUS_ID format.

For example, an event ARN might look like the following:

arn:aws:health:us-east-1::event/EC2/EC2_INSTANCE_RETIREMENT_SCHEDULED/EC2_INSTANCE_RETIREMENT_SCHEDULED_ABC123-DEF456

organizationEvent_service :: Lens' OrganizationEvent (Maybe Text)

The AWS service that is affected by the event, such as EC2 and RDS.

organizationEvent_startTime :: Lens' OrganizationEvent (Maybe UTCTime)

The date and time that the event began.

organizationEvent_eventScopeCode :: Lens' OrganizationEvent (Maybe EventScopeCode)

This parameter specifies if the AWS Health event is a public AWS service event or an account-specific event.

If the eventScopeCode value is PUBLIC, then the affectedAccounts value is always empty.

If the eventScopeCode value is ACCOUNT_SPECIFIC, then the affectedAccounts value lists the affected AWS accounts in your organization. For example, if an event affects a service such as Amazon Elastic Compute Cloud and you have AWS accounts that use that service, those account IDs appear in the response.

If the eventScopeCode value is NONE, then the eventArn that you specified in the request is invalid or doesn't exist.

organizationEvent_eventTypeCode :: Lens' OrganizationEvent (Maybe Text)

The unique identifier for the event type. The format is AWS_SERVICE_DESCRIPTION. For example, AWS_EC2_SYSTEM_MAINTENANCE_EVENT.

organizationEvent_eventTypeCategory :: Lens' OrganizationEvent (Maybe EventTypeCategory)

The category of the event type.

organizationEvent_endTime :: Lens' OrganizationEvent (Maybe UTCTime)

The date and time that the event ended.

organizationEvent_region :: Lens' OrganizationEvent (Maybe Text)

The AWS Region name of the event.

organizationEvent_statusCode :: Lens' OrganizationEvent (Maybe EventStatusCode)

The most recent status of the event. Possible values are open, closed, and upcoming.