Network.AWS.AccessAnalyzer.CancelPolicyGeneration

Description

Cancels the requested policy generation.

Creating a Request

data CancelPolicyGeneration

See: newCancelPolicyGeneration smart constructor.

Constructors

CancelPolicyGeneration'

Fields

jobId :: Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

newCancelPolicyGeneration

Arguments

:: Text
$sel:jobId:CancelPolicyGeneration'

-> CancelPolicyGeneration

Create a value of CancelPolicyGeneration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:CancelPolicyGeneration', cancelPolicyGeneration_jobId - The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

Request Lenses

cancelPolicyGeneration_jobId :: Lens' CancelPolicyGeneration Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

Destructuring the Response

data CancelPolicyGenerationResponse

See: newCancelPolicyGenerationResponse smart constructor.

Constructors

CancelPolicyGenerationResponse'

Fields

httpStatus :: Int

The response's http status code.

newCancelPolicyGenerationResponse

Arguments

:: Int
$sel:httpStatus:CancelPolicyGenerationResponse'

-> CancelPolicyGenerationResponse

Create a value of CancelPolicyGenerationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CancelPolicyGenerationResponse', cancelPolicyGenerationResponse_httpStatus - The response's http status code.

Response Lenses

cancelPolicyGenerationResponse_httpStatus :: Lens' CancelPolicyGenerationResponse Int

The response's http status code.