Network.AWS.AccessAnalyzer.CreateAnalyzer

Description

Creates an analyzer for your account.

Creating a Request

data CreateAnalyzer

Creates an analyzer.

See: newCreateAnalyzer smart constructor.

Constructors

CreateAnalyzer'

Fields

clientToken :: Maybe Text

A client token.

archiveRules :: Maybe [InlineArchiveRule]

Specifies the archive rules to add for the analyzer. Archive rules automatically archive findings that meet the criteria you define for the rule.

tags :: Maybe (HashMap Text Text)

The tags to apply to the analyzer.

analyzerName :: Text

The name of the analyzer to create.

type' :: Type

The type of analyzer to create. Only ACCOUNT and ORGANIZATION analyzers are supported. You can create only one analyzer per account per Region. You can create up to 5 analyzers per organization per Region.

newCreateAnalyzer

Arguments

:: Text
$sel:analyzerName:CreateAnalyzer'

-> Type
$sel:type':CreateAnalyzer'

-> CreateAnalyzer

Create a value of CreateAnalyzer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateAnalyzer', createAnalyzer_clientToken - A client token.

$sel:archiveRules:CreateAnalyzer', createAnalyzer_archiveRules - Specifies the archive rules to add for the analyzer. Archive rules automatically archive findings that meet the criteria you define for the rule.

$sel:tags:CreateAnalyzer', createAnalyzer_tags - The tags to apply to the analyzer.

$sel:analyzerName:CreateAnalyzer', createAnalyzer_analyzerName - The name of the analyzer to create.

$sel:type':CreateAnalyzer', createAnalyzer_type - The type of analyzer to create. Only ACCOUNT and ORGANIZATION analyzers are supported. You can create only one analyzer per account per Region. You can create up to 5 analyzers per organization per Region.

Request Lenses

createAnalyzer_clientToken :: Lens' CreateAnalyzer (Maybe Text)

A client token.

createAnalyzer_archiveRules :: Lens' CreateAnalyzer (Maybe [InlineArchiveRule])

Specifies the archive rules to add for the analyzer. Archive rules automatically archive findings that meet the criteria you define for the rule.

createAnalyzer_tags :: Lens' CreateAnalyzer (Maybe (HashMap Text Text))

The tags to apply to the analyzer.

createAnalyzer_analyzerName :: Lens' CreateAnalyzer Text

The name of the analyzer to create.

createAnalyzer_type :: Lens' CreateAnalyzer Type

The type of analyzer to create. Only ACCOUNT and ORGANIZATION analyzers are supported. You can create only one analyzer per account per Region. You can create up to 5 analyzers per organization per Region.

Destructuring the Response

data CreateAnalyzerResponse

The response to the request to create an analyzer.

See: newCreateAnalyzerResponse smart constructor.

Constructors

CreateAnalyzerResponse'

Fields

arn :: Maybe Text

The ARN of the analyzer that was created by the request.

httpStatus :: Int

The response's http status code.

newCreateAnalyzerResponse

Arguments

:: Int
$sel:httpStatus:CreateAnalyzerResponse'

-> CreateAnalyzerResponse

Create a value of CreateAnalyzerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateAnalyzerResponse', createAnalyzerResponse_arn - The ARN of the analyzer that was created by the request.

$sel:httpStatus:CreateAnalyzerResponse', createAnalyzerResponse_httpStatus - The response's http status code.

Response Lenses

createAnalyzerResponse_arn :: Lens' CreateAnalyzerResponse (Maybe Text)

The ARN of the analyzer that was created by the request.

createAnalyzerResponse_httpStatus :: Lens' CreateAnalyzerResponse Int

The response's http status code.