Network.AWS.AccessAnalyzer.GetAnalyzedResource

Description

Retrieves information about a resource that was analyzed.

Creating a Request

data GetAnalyzedResource

Retrieves an analyzed resource.

See: newGetAnalyzedResource smart constructor.

Constructors

GetAnalyzedResource'

Fields

analyzerArn :: Text

The ARN of the analyzer to retrieve information from.

resourceArn :: Text

The ARN of the resource to retrieve information about.

newGetAnalyzedResource

Arguments

:: Text
$sel:analyzerArn:GetAnalyzedResource'

-> Text
$sel:resourceArn:GetAnalyzedResource'

-> GetAnalyzedResource

Create a value of GetAnalyzedResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:analyzerArn:GetAnalyzedResource', getAnalyzedResource_analyzerArn - The ARN of the analyzer to retrieve information from.

$sel:resourceArn:GetAnalyzedResource', getAnalyzedResource_resourceArn - The ARN of the resource to retrieve information about.

Request Lenses

getAnalyzedResource_analyzerArn :: Lens' GetAnalyzedResource Text

The ARN of the analyzer to retrieve information from.

getAnalyzedResource_resourceArn :: Lens' GetAnalyzedResource Text

The ARN of the resource to retrieve information about.

Destructuring the Response

data GetAnalyzedResourceResponse

The response to the request.

See: newGetAnalyzedResourceResponse smart constructor.

Constructors

GetAnalyzedResourceResponse'

Fields

resource :: Maybe AnalyzedResource

An AnalyzedResource object that contains information that IAM Access Analyzer found when it analyzed the resource.

httpStatus :: Int

The response's http status code.

newGetAnalyzedResourceResponse

Arguments

:: Int
$sel:httpStatus:GetAnalyzedResourceResponse'

-> GetAnalyzedResourceResponse

Create a value of GetAnalyzedResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resource:GetAnalyzedResourceResponse', getAnalyzedResourceResponse_resource - An AnalyzedResource object that contains information that IAM Access Analyzer found when it analyzed the resource.

$sel:httpStatus:GetAnalyzedResourceResponse', getAnalyzedResourceResponse_httpStatus - The response's http status code.

Response Lenses

getAnalyzedResourceResponse_resource :: Lens' GetAnalyzedResourceResponse (Maybe AnalyzedResource)

An AnalyzedResource object that contains information that IAM Access Analyzer found when it analyzed the resource.

getAnalyzedResourceResponse_httpStatus :: Lens' GetAnalyzedResourceResponse Int

The response's http status code.