Network.AWS.AccessAnalyzer.GetGeneratedPolicy

Description

Retrieves the policy that was generated using StartPolicyGeneration.

Creating a Request

data GetGeneratedPolicy

See: newGetGeneratedPolicy smart constructor.

Constructors

GetGeneratedPolicy'

Fields

includeServiceLevelTemplate :: Maybe Bool

The level of detail that you want to generate. You can specify whether to generate service-level policies.

IAM Access Analyzer uses iam:servicelastaccessed to identify services that have been used recently to create this service-level template.

includeResourcePlaceholders :: Maybe Bool

The level of detail that you want to generate. You can specify whether to generate policies with placeholders for resource ARNs for actions that support resource level granularity in policies.

For example, in the resource section of a policy, you can receive a placeholder such as "Resource":"arn:aws:s3:::${BucketName}" instead of "*".

jobId :: Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

newGetGeneratedPolicy

Arguments

:: Text
$sel:jobId:GetGeneratedPolicy'

-> GetGeneratedPolicy

Create a value of GetGeneratedPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeServiceLevelTemplate:GetGeneratedPolicy', getGeneratedPolicy_includeServiceLevelTemplate - The level of detail that you want to generate. You can specify whether to generate service-level policies.

IAM Access Analyzer uses iam:servicelastaccessed to identify services that have been used recently to create this service-level template.

$sel:includeResourcePlaceholders:GetGeneratedPolicy', getGeneratedPolicy_includeResourcePlaceholders - The level of detail that you want to generate. You can specify whether to generate policies with placeholders for resource ARNs for actions that support resource level granularity in policies.

For example, in the resource section of a policy, you can receive a placeholder such as "Resource":"arn:aws:s3:::${BucketName}" instead of "*".

$sel:jobId:GetGeneratedPolicy', getGeneratedPolicy_jobId - The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

Request Lenses

getGeneratedPolicy_includeServiceLevelTemplate :: Lens' GetGeneratedPolicy (Maybe Bool)

The level of detail that you want to generate. You can specify whether to generate service-level policies.

IAM Access Analyzer uses iam:servicelastaccessed to identify services that have been used recently to create this service-level template.

getGeneratedPolicy_includeResourcePlaceholders :: Lens' GetGeneratedPolicy (Maybe Bool)

The level of detail that you want to generate. You can specify whether to generate policies with placeholders for resource ARNs for actions that support resource level granularity in policies.

For example, in the resource section of a policy, you can receive a placeholder such as "Resource":"arn:aws:s3:::${BucketName}" instead of "*".

getGeneratedPolicy_jobId :: Lens' GetGeneratedPolicy Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

Destructuring the Response

data GetGeneratedPolicyResponse

See: newGetGeneratedPolicyResponse smart constructor.

Constructors

GetGeneratedPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

generatedPolicyResult :: GeneratedPolicyResult

A GeneratedPolicyResult object that contains the generated policies and associated details.

jobDetails :: JobDetails

A GeneratedPolicyDetails object that contains details about the generated policy.

newGetGeneratedPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetGeneratedPolicyResponse'

-> GeneratedPolicyResult
$sel:generatedPolicyResult:GetGeneratedPolicyResponse'

-> JobDetails
$sel:jobDetails:GetGeneratedPolicyResponse'

-> GetGeneratedPolicyResponse

Create a value of GetGeneratedPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetGeneratedPolicyResponse', getGeneratedPolicyResponse_httpStatus - The response's http status code.

$sel:generatedPolicyResult:GetGeneratedPolicyResponse', getGeneratedPolicyResponse_generatedPolicyResult - A GeneratedPolicyResult object that contains the generated policies and associated details.

$sel:jobDetails:GetGeneratedPolicyResponse', getGeneratedPolicyResponse_jobDetails - A GeneratedPolicyDetails object that contains details about the generated policy.

Response Lenses

getGeneratedPolicyResponse_httpStatus :: Lens' GetGeneratedPolicyResponse Int

The response's http status code.

getGeneratedPolicyResponse_generatedPolicyResult :: Lens' GetGeneratedPolicyResponse GeneratedPolicyResult

A GeneratedPolicyResult object that contains the generated policies and associated details.

getGeneratedPolicyResponse_jobDetails :: Lens' GetGeneratedPolicyResponse JobDetails

A GeneratedPolicyDetails object that contains details about the generated policy.