Network.AWS.AccessAnalyzer.ListAnalyzedResources

Description

Retrieves a list of resources of the specified type that have been analyzed by the specified analyzer..

This operation returns paginated results.

Creating a Request

data ListAnalyzedResources

Retrieves a list of resources that have been analyzed.

See: newListAnalyzedResources smart constructor.

Constructors

ListAnalyzedResources'

Fields

resourceType :: Maybe ResourceType

The type of resource.

nextToken :: Maybe Text

A token used for pagination of results returned.

maxResults :: Maybe Int

The maximum number of results to return in the response.

analyzerArn :: Text

The ARN of the analyzer to retrieve a list of analyzed resources from.

newListAnalyzedResources

Arguments

:: Text
$sel:analyzerArn:ListAnalyzedResources'

-> ListAnalyzedResources

Create a value of ListAnalyzedResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ListAnalyzedResources', listAnalyzedResources_resourceType - The type of resource.

$sel:nextToken:ListAnalyzedResources', listAnalyzedResources_nextToken - A token used for pagination of results returned.

$sel:maxResults:ListAnalyzedResources', listAnalyzedResources_maxResults - The maximum number of results to return in the response.

$sel:analyzerArn:ListAnalyzedResources', listAnalyzedResources_analyzerArn - The ARN of the analyzer to retrieve a list of analyzed resources from.

Request Lenses

listAnalyzedResources_resourceType :: Lens' ListAnalyzedResources (Maybe ResourceType)

The type of resource.

listAnalyzedResources_nextToken :: Lens' ListAnalyzedResources (Maybe Text)

A token used for pagination of results returned.

listAnalyzedResources_maxResults :: Lens' ListAnalyzedResources (Maybe Int)

The maximum number of results to return in the response.

listAnalyzedResources_analyzerArn :: Lens' ListAnalyzedResources Text

The ARN of the analyzer to retrieve a list of analyzed resources from.

Destructuring the Response

data ListAnalyzedResourcesResponse

The response to the request.

See: newListAnalyzedResourcesResponse smart constructor.

Constructors

ListAnalyzedResourcesResponse'

Fields

nextToken :: Maybe Text

A token used for pagination of results returned.

httpStatus :: Int

The response's http status code.

analyzedResources :: [AnalyzedResourceSummary]

A list of resources that were analyzed.

newListAnalyzedResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListAnalyzedResourcesResponse'

-> ListAnalyzedResourcesResponse

Create a value of ListAnalyzedResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAnalyzedResources', listAnalyzedResourcesResponse_nextToken - A token used for pagination of results returned.

$sel:httpStatus:ListAnalyzedResourcesResponse', listAnalyzedResourcesResponse_httpStatus - The response's http status code.

$sel:analyzedResources:ListAnalyzedResourcesResponse', listAnalyzedResourcesResponse_analyzedResources - A list of resources that were analyzed.

Response Lenses

listAnalyzedResourcesResponse_nextToken :: Lens' ListAnalyzedResourcesResponse (Maybe Text)

A token used for pagination of results returned.

listAnalyzedResourcesResponse_httpStatus :: Lens' ListAnalyzedResourcesResponse Int

The response's http status code.

listAnalyzedResourcesResponse_analyzedResources :: Lens' ListAnalyzedResourcesResponse [AnalyzedResourceSummary]

A list of resources that were analyzed.