Network.AWS.AccessAnalyzer.ListAnalyzers

Description

Retrieves a list of analyzers.

This operation returns paginated results.

Creating a Request

data ListAnalyzers

Retrieves a list of analyzers.

See: newListAnalyzers smart constructor.

Constructors

ListAnalyzers'

Fields

nextToken :: Maybe Text

A token used for pagination of results returned.

type' :: Maybe Type

The type of analyzer.

maxResults :: Maybe Int

The maximum number of results to return in the response.

newListAnalyzers :: ListAnalyzers

Create a value of ListAnalyzers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAnalyzers', listAnalyzers_nextToken - A token used for pagination of results returned.

$sel:type':ListAnalyzers', listAnalyzers_type - The type of analyzer.

$sel:maxResults:ListAnalyzers', listAnalyzers_maxResults - The maximum number of results to return in the response.

Request Lenses

listAnalyzers_nextToken :: Lens' ListAnalyzers (Maybe Text)

A token used for pagination of results returned.

listAnalyzers_type :: Lens' ListAnalyzers (Maybe Type)

The type of analyzer.

listAnalyzers_maxResults :: Lens' ListAnalyzers (Maybe Int)

The maximum number of results to return in the response.

Destructuring the Response

data ListAnalyzersResponse

The response to the request.

See: newListAnalyzersResponse smart constructor.

Constructors

ListAnalyzersResponse'

Fields

nextToken :: Maybe Text

A token used for pagination of results returned.

httpStatus :: Int

The response's http status code.

analyzers :: [AnalyzerSummary]

The analyzers retrieved.

newListAnalyzersResponse

Arguments

:: Int
$sel:httpStatus:ListAnalyzersResponse'

-> ListAnalyzersResponse

Create a value of ListAnalyzersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAnalyzers', listAnalyzersResponse_nextToken - A token used for pagination of results returned.

$sel:httpStatus:ListAnalyzersResponse', listAnalyzersResponse_httpStatus - The response's http status code.

$sel:analyzers:ListAnalyzersResponse', listAnalyzersResponse_analyzers - The analyzers retrieved.

Response Lenses

listAnalyzersResponse_nextToken :: Lens' ListAnalyzersResponse (Maybe Text)

A token used for pagination of results returned.

listAnalyzersResponse_httpStatus :: Lens' ListAnalyzersResponse Int

The response's http status code.

listAnalyzersResponse_analyzers :: Lens' ListAnalyzersResponse [AnalyzerSummary]

The analyzers retrieved.