Network.AWS.AccessAnalyzer.Types.AccessPreviewStatusReason

Description

Documentation

data AccessPreviewStatusReason

Provides more details about the current status of the access preview. For example, if the creation of the access preview fails, a Failed status is returned. This failure can be due to an internal issue with the analysis or due to an invalid proposed resource configuration.

See: newAccessPreviewStatusReason smart constructor.

Constructors

AccessPreviewStatusReason'

Fields

code :: AccessPreviewStatusReasonCode

The reason code for the current status of the access preview.

newAccessPreviewStatusReason

Arguments

:: AccessPreviewStatusReasonCode
$sel:code:AccessPreviewStatusReason'

-> AccessPreviewStatusReason

Create a value of AccessPreviewStatusReason with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:code:AccessPreviewStatusReason', accessPreviewStatusReason_code - The reason code for the current status of the access preview.

accessPreviewStatusReason_code :: Lens' AccessPreviewStatusReason AccessPreviewStatusReasonCode

The reason code for the current status of the access preview.