Network.AWS.AccessAnalyzer.Types.AnalyzerSummary

Description

Documentation

data AnalyzerSummary

Contains information about the analyzer.

See: newAnalyzerSummary smart constructor.

Constructors

AnalyzerSummary'

Fields

lastResourceAnalyzedAt :: Maybe POSIX

The time at which the most recently analyzed resource was analyzed.

lastResourceAnalyzed :: Maybe Text

The resource that was most recently analyzed by the analyzer.

statusReason :: Maybe StatusReason

The statusReason provides more details about the current status of the analyzer. For example, if the creation for the analyzer fails, a Failed status is returned. For an analyzer with organization as the type, this failure can be due to an issue with creating the service-linked roles required in the member accounts of the Amazon Web Services organization.

tags :: Maybe (HashMap Text Text)

The tags added to the analyzer.

arn :: Text

The ARN of the analyzer.

createdAt :: POSIX

A timestamp for the time at which the analyzer was created.

name :: Text

The name of the analyzer.

status :: AnalyzerStatus

The status of the analyzer. An Active analyzer successfully monitors supported resources and generates new findings. The analyzer is Disabled when a user action, such as removing trusted access for Identity and Access Management Access Analyzer from Organizations, causes the analyzer to stop generating new findings. The status is Creating when the analyzer creation is in progress and Failed when the analyzer creation has failed.

type' :: Type

The type of analyzer, which corresponds to the zone of trust chosen for the analyzer.

newAnalyzerSummary

Arguments

:: Text
$sel:arn:AnalyzerSummary'

-> UTCTime
$sel:createdAt:AnalyzerSummary'

-> Text
$sel:name:AnalyzerSummary'

-> AnalyzerStatus
$sel:status:AnalyzerSummary'

-> Type
$sel:type':AnalyzerSummary'

-> AnalyzerSummary

Create a value of AnalyzerSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastResourceAnalyzedAt:AnalyzerSummary', analyzerSummary_lastResourceAnalyzedAt - The time at which the most recently analyzed resource was analyzed.

$sel:lastResourceAnalyzed:AnalyzerSummary', analyzerSummary_lastResourceAnalyzed - The resource that was most recently analyzed by the analyzer.

$sel:statusReason:AnalyzerSummary', analyzerSummary_statusReason - The statusReason provides more details about the current status of the analyzer. For example, if the creation for the analyzer fails, a Failed status is returned. For an analyzer with organization as the type, this failure can be due to an issue with creating the service-linked roles required in the member accounts of the Amazon Web Services organization.

$sel:tags:AnalyzerSummary', analyzerSummary_tags - The tags added to the analyzer.

$sel:arn:AnalyzerSummary', analyzerSummary_arn - The ARN of the analyzer.

$sel:createdAt:AnalyzerSummary', analyzerSummary_createdAt - A timestamp for the time at which the analyzer was created.

$sel:name:AnalyzerSummary', analyzerSummary_name - The name of the analyzer.

$sel:status:AnalyzerSummary', analyzerSummary_status - The status of the analyzer. An Active analyzer successfully monitors supported resources and generates new findings. The analyzer is Disabled when a user action, such as removing trusted access for Identity and Access Management Access Analyzer from Organizations, causes the analyzer to stop generating new findings. The status is Creating when the analyzer creation is in progress and Failed when the analyzer creation has failed.

$sel:type':AnalyzerSummary', analyzerSummary_type - The type of analyzer, which corresponds to the zone of trust chosen for the analyzer.

analyzerSummary_lastResourceAnalyzedAt :: Lens' AnalyzerSummary (Maybe UTCTime)

The time at which the most recently analyzed resource was analyzed.

analyzerSummary_lastResourceAnalyzed :: Lens' AnalyzerSummary (Maybe Text)

The resource that was most recently analyzed by the analyzer.

analyzerSummary_statusReason :: Lens' AnalyzerSummary (Maybe StatusReason)

The statusReason provides more details about the current status of the analyzer. For example, if the creation for the analyzer fails, a Failed status is returned. For an analyzer with organization as the type, this failure can be due to an issue with creating the service-linked roles required in the member accounts of the Amazon Web Services organization.

analyzerSummary_tags :: Lens' AnalyzerSummary (Maybe (HashMap Text Text))

The tags added to the analyzer.

analyzerSummary_arn :: Lens' AnalyzerSummary Text

The ARN of the analyzer.

analyzerSummary_createdAt :: Lens' AnalyzerSummary UTCTime

A timestamp for the time at which the analyzer was created.

analyzerSummary_name :: Lens' AnalyzerSummary Text

The name of the analyzer.

analyzerSummary_status :: Lens' AnalyzerSummary AnalyzerStatus

The status of the analyzer. An Active analyzer successfully monitors supported resources and generates new findings. The analyzer is Disabled when a user action, such as removing trusted access for Identity and Access Management Access Analyzer from Organizations, causes the analyzer to stop generating new findings. The status is Creating when the analyzer creation is in progress and Failed when the analyzer creation has failed.

analyzerSummary_type :: Lens' AnalyzerSummary Type

The type of analyzer, which corresponds to the zone of trust chosen for the analyzer.