Network.AWS.AccessAnalyzer.Types.CloudTrailProperties

Description

Documentation

data CloudTrailProperties

Contains information about CloudTrail access.

See: newCloudTrailProperties smart constructor.

Constructors

CloudTrailProperties'

Fields

endTime :: POSIX

The end of the time range for which IAM Access Analyzer reviews your CloudTrail events. Events with a timestamp after this time are not considered to generate a policy. If this is not included in the request, the default value is the current time.

startTime :: POSIX

The start of the time range for which IAM Access Analyzer reviews your CloudTrail events. Events with a timestamp before this time are not considered to generate a policy.

trailProperties :: [TrailProperties]

A TrailProperties object that contains settings for trail properties.

newCloudTrailProperties

Arguments

:: UTCTime
$sel:endTime:CloudTrailProperties'

-> UTCTime
$sel:startTime:CloudTrailProperties'

-> CloudTrailProperties

Create a value of CloudTrailProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endTime:CloudTrailProperties', cloudTrailProperties_endTime - The end of the time range for which IAM Access Analyzer reviews your CloudTrail events. Events with a timestamp after this time are not considered to generate a policy. If this is not included in the request, the default value is the current time.

$sel:startTime:CloudTrailProperties', cloudTrailProperties_startTime - The start of the time range for which IAM Access Analyzer reviews your CloudTrail events. Events with a timestamp before this time are not considered to generate a policy.

$sel:trailProperties:CloudTrailProperties', cloudTrailProperties_trailProperties - A TrailProperties object that contains settings for trail properties.

cloudTrailProperties_endTime :: Lens' CloudTrailProperties UTCTime

The end of the time range for which IAM Access Analyzer reviews your CloudTrail events. Events with a timestamp after this time are not considered to generate a policy. If this is not included in the request, the default value is the current time.

cloudTrailProperties_startTime :: Lens' CloudTrailProperties UTCTime

The start of the time range for which IAM Access Analyzer reviews your CloudTrail events. Events with a timestamp before this time are not considered to generate a policy.

cloudTrailProperties_trailProperties :: Lens' CloudTrailProperties [TrailProperties]

A TrailProperties object that contains settings for trail properties.