Network.AWS.AccessAnalyzer.Types.Configuration

Description

Documentation

data Configuration

Access control configuration structures for your resource. You specify the configuration as a type-value pair. You can specify only one type of access control configuration.

See: newConfiguration smart constructor.

Constructors

Configuration'

Fields

kmsKey :: Maybe KmsKeyConfiguration

The access control configuration is for a KMS key.

secretsManagerSecret :: Maybe SecretsManagerSecretConfiguration

The access control configuration is for a Secrets Manager secret.

sqsQueue :: Maybe SqsQueueConfiguration

The access control configuration is for an Amazon SQS queue.

s3Bucket :: Maybe S3BucketConfiguration

The access control configuration is for an Amazon S3 Bucket.

iamRole :: Maybe IamRoleConfiguration

The access control configuration is for an IAM role.

newConfiguration :: Configuration

Create a value of Configuration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKey:Configuration', configuration_kmsKey - The access control configuration is for a KMS key.

$sel:secretsManagerSecret:Configuration', configuration_secretsManagerSecret - The access control configuration is for a Secrets Manager secret.

$sel:sqsQueue:Configuration', configuration_sqsQueue - The access control configuration is for an Amazon SQS queue.

$sel:s3Bucket:Configuration', configuration_s3Bucket - The access control configuration is for an Amazon S3 Bucket.

$sel:iamRole:Configuration', configuration_iamRole - The access control configuration is for an IAM role.

configuration_kmsKey :: Lens' Configuration (Maybe KmsKeyConfiguration)

The access control configuration is for a KMS key.

configuration_secretsManagerSecret :: Lens' Configuration (Maybe SecretsManagerSecretConfiguration)

The access control configuration is for a Secrets Manager secret.

configuration_sqsQueue :: Lens' Configuration (Maybe SqsQueueConfiguration)

The access control configuration is for an Amazon SQS queue.

configuration_s3Bucket :: Lens' Configuration (Maybe S3BucketConfiguration)

The access control configuration is for an Amazon S3 Bucket.

configuration_iamRole :: Lens' Configuration (Maybe IamRoleConfiguration)

The access control configuration is for an IAM role.