Network.AWS.AccessAnalyzer.Types.FindingSummary

Description

Documentation

data FindingSummary

Contains information about a finding.

See: newFindingSummary smart constructor.

Constructors

FindingSummary'

Fields

error :: Maybe Text

The error that resulted in an Error finding.

isPublic :: Maybe Bool

Indicates whether the finding reports a resource that has a policy that allows public access.

action :: Maybe [Text]

The action in the analyzed policy statement that an external principal has permission to use.

sources :: Maybe [FindingSource]

The sources of the finding. This indicates how the access that generated the finding is granted. It is populated for Amazon S3 bucket findings.

resource :: Maybe Text

The resource that the external principal has access to.

principal :: Maybe (HashMap Text Text)

The external principal that has access to a resource within the zone of trust.

analyzedAt :: POSIX

The time at which the resource-based policy that generated the finding was analyzed.

condition :: HashMap Text Text

The condition in the analyzed policy statement that resulted in a finding.

createdAt :: POSIX

The time at which the finding was created.

id :: Text

The ID of the finding.

resourceOwnerAccount :: Text

The Amazon Web Services account ID that owns the resource.

resourceType :: ResourceType

The type of the resource that the external principal has access to.

status :: FindingStatus

The status of the finding.

updatedAt :: POSIX

The time at which the finding was most recently updated.

newFindingSummary

Arguments

:: UTCTime
$sel:analyzedAt:FindingSummary'

-> UTCTime
$sel:createdAt:FindingSummary'

-> Text
$sel:id:FindingSummary'

-> Text
$sel:resourceOwnerAccount:FindingSummary'

-> ResourceType
$sel:resourceType:FindingSummary'

-> FindingStatus
$sel:status:FindingSummary'

-> UTCTime
$sel:updatedAt:FindingSummary'

-> FindingSummary

Create a value of FindingSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:error:FindingSummary', findingSummary_error - The error that resulted in an Error finding.

$sel:isPublic:FindingSummary', findingSummary_isPublic - Indicates whether the finding reports a resource that has a policy that allows public access.

$sel:action:FindingSummary', findingSummary_action - The action in the analyzed policy statement that an external principal has permission to use.

$sel:sources:FindingSummary', findingSummary_sources - The sources of the finding. This indicates how the access that generated the finding is granted. It is populated for Amazon S3 bucket findings.

$sel:resource:FindingSummary', findingSummary_resource - The resource that the external principal has access to.

$sel:principal:FindingSummary', findingSummary_principal - The external principal that has access to a resource within the zone of trust.

$sel:analyzedAt:FindingSummary', findingSummary_analyzedAt - The time at which the resource-based policy that generated the finding was analyzed.

$sel:condition:FindingSummary', findingSummary_condition - The condition in the analyzed policy statement that resulted in a finding.

$sel:createdAt:FindingSummary', findingSummary_createdAt - The time at which the finding was created.

$sel:id:FindingSummary', findingSummary_id - The ID of the finding.

$sel:resourceOwnerAccount:FindingSummary', findingSummary_resourceOwnerAccount - The Amazon Web Services account ID that owns the resource.

$sel:resourceType:FindingSummary', findingSummary_resourceType - The type of the resource that the external principal has access to.

$sel:status:FindingSummary', findingSummary_status - The status of the finding.

$sel:updatedAt:FindingSummary', findingSummary_updatedAt - The time at which the finding was most recently updated.

findingSummary_error :: Lens' FindingSummary (Maybe Text)

The error that resulted in an Error finding.

findingSummary_isPublic :: Lens' FindingSummary (Maybe Bool)

Indicates whether the finding reports a resource that has a policy that allows public access.

findingSummary_action :: Lens' FindingSummary (Maybe [Text])

The action in the analyzed policy statement that an external principal has permission to use.

findingSummary_sources :: Lens' FindingSummary (Maybe [FindingSource])

The sources of the finding. This indicates how the access that generated the finding is granted. It is populated for Amazon S3 bucket findings.

findingSummary_resource :: Lens' FindingSummary (Maybe Text)

The resource that the external principal has access to.

findingSummary_principal :: Lens' FindingSummary (Maybe (HashMap Text Text))

The external principal that has access to a resource within the zone of trust.

findingSummary_analyzedAt :: Lens' FindingSummary UTCTime

The time at which the resource-based policy that generated the finding was analyzed.

findingSummary_condition :: Lens' FindingSummary (HashMap Text Text)

The condition in the analyzed policy statement that resulted in a finding.

findingSummary_createdAt :: Lens' FindingSummary UTCTime

The time at which the finding was created.

findingSummary_id :: Lens' FindingSummary Text

The ID of the finding.

findingSummary_resourceOwnerAccount :: Lens' FindingSummary Text

The Amazon Web Services account ID that owns the resource.

findingSummary_resourceType :: Lens' FindingSummary ResourceType

The type of the resource that the external principal has access to.

findingSummary_status :: Lens' FindingSummary FindingStatus

The status of the finding.

findingSummary_updatedAt :: Lens' FindingSummary UTCTime

The time at which the finding was most recently updated.