Network.AWS.AccessAnalyzer.Types.GeneratedPolicyResult

Description

Documentation

data GeneratedPolicyResult

Contains the text for the generated policy and its details.

See: newGeneratedPolicyResult smart constructor.

Constructors

GeneratedPolicyResult'

Fields

generatedPolicies :: Maybe [GeneratedPolicy]

The text to use as the content for the new policy. The policy is created using the CreatePolicy action.

properties :: GeneratedPolicyProperties

A GeneratedPolicyProperties object that contains properties of the generated policy.

newGeneratedPolicyResult

Arguments

:: GeneratedPolicyProperties
$sel:properties:GeneratedPolicyResult'

-> GeneratedPolicyResult

Create a value of GeneratedPolicyResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:generatedPolicies:GeneratedPolicyResult', generatedPolicyResult_generatedPolicies - The text to use as the content for the new policy. The policy is created using the CreatePolicy action.

$sel:properties:GeneratedPolicyResult', generatedPolicyResult_properties - A GeneratedPolicyProperties object that contains properties of the generated policy.

generatedPolicyResult_generatedPolicies :: Lens' GeneratedPolicyResult (Maybe [GeneratedPolicy])

The text to use as the content for the new policy. The policy is created using the CreatePolicy action.

generatedPolicyResult_properties :: Lens' GeneratedPolicyResult GeneratedPolicyProperties

A GeneratedPolicyProperties object that contains properties of the generated policy.