Network.AWS.AccessAnalyzer.Types.JobDetails

Description

Documentation

data JobDetails

Contains details about the policy generation request.

See: newJobDetails smart constructor.

Constructors

JobDetails'

Fields

completedOn :: Maybe POSIX

A timestamp of when the job was completed.

jobError :: Maybe JobError

The job error for the policy generation request.

jobId :: Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

startedOn :: POSIX

A timestamp of when the job was started.

status :: JobStatus

The status of the job request.

newJobDetails

Arguments

:: Text
$sel:jobId:JobDetails'

-> UTCTime
$sel:startedOn:JobDetails'

-> JobStatus
$sel:status:JobDetails'

-> JobDetails

Create a value of JobDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:completedOn:JobDetails', jobDetails_completedOn - A timestamp of when the job was completed.

$sel:jobError:JobDetails', jobDetails_jobError - The job error for the policy generation request.

$sel:jobId:JobDetails', jobDetails_jobId - The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

$sel:startedOn:JobDetails', jobDetails_startedOn - A timestamp of when the job was started.

$sel:status:JobDetails', jobDetails_status - The status of the job request.

jobDetails_completedOn :: Lens' JobDetails (Maybe UTCTime)

A timestamp of when the job was completed.

jobDetails_jobError :: Lens' JobDetails (Maybe JobError)

The job error for the policy generation request.

jobDetails_jobId :: Lens' JobDetails Text

The JobId that is returned by the StartPolicyGeneration operation. The JobId can be used with GetGeneratedPolicy to retrieve the generated policies or used with CancelPolicyGeneration to cancel the policy generation request.

jobDetails_startedOn :: Lens' JobDetails UTCTime

A timestamp of when the job was started.

jobDetails_status :: Lens' JobDetails JobStatus

The status of the job request.