Network.AWS.AccessAnalyzer.Types.S3AccessPointConfiguration

Description

Documentation

data S3AccessPointConfiguration

The configuration for an Amazon S3 access point or multi-region access point for the bucket. You can propose up to 10 access points or multi-region access points per bucket. If the proposed Amazon S3 access point configuration is for an existing bucket, the access preview uses the proposed access point configuration in place of the existing access points. To propose an access point without a policy, you can provide an empty string as the access point policy. For more information, see Creating access points. For more information about access point policy limits, see Access points restrictions and limitations.

See: newS3AccessPointConfiguration smart constructor.

Constructors

S3AccessPointConfiguration'

Fields

publicAccessBlock :: Maybe S3PublicAccessBlockConfiguration

The proposed S3PublicAccessBlock configuration to apply to this Amazon S3 access point or multi-region access point.

accessPointPolicy :: Maybe Text

The access point or multi-region access point policy.

networkOrigin :: Maybe NetworkOriginConfiguration

The proposed Internet and VpcConfiguration to apply to this Amazon S3 access point. VpcConfiguration does not apply to multi-region access points. If the access preview is for a new resource and neither is specified, the access preview uses Internet for the network origin. If the access preview is for an existing resource and neither is specified, the access preview uses the exiting network origin.

newS3AccessPointConfiguration :: S3AccessPointConfiguration

Create a value of S3AccessPointConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:publicAccessBlock:S3AccessPointConfiguration', s3AccessPointConfiguration_publicAccessBlock - The proposed S3PublicAccessBlock configuration to apply to this Amazon S3 access point or multi-region access point.

$sel:accessPointPolicy:S3AccessPointConfiguration', s3AccessPointConfiguration_accessPointPolicy - The access point or multi-region access point policy.

$sel:networkOrigin:S3AccessPointConfiguration', s3AccessPointConfiguration_networkOrigin - The proposed Internet and VpcConfiguration to apply to this Amazon S3 access point. VpcConfiguration does not apply to multi-region access points. If the access preview is for a new resource and neither is specified, the access preview uses Internet for the network origin. If the access preview is for an existing resource and neither is specified, the access preview uses the exiting network origin.

s3AccessPointConfiguration_publicAccessBlock :: Lens' S3AccessPointConfiguration (Maybe S3PublicAccessBlockConfiguration)

The proposed S3PublicAccessBlock configuration to apply to this Amazon S3 access point or multi-region access point.

s3AccessPointConfiguration_accessPointPolicy :: Lens' S3AccessPointConfiguration (Maybe Text)

The access point or multi-region access point policy.

s3AccessPointConfiguration_networkOrigin :: Lens' S3AccessPointConfiguration (Maybe NetworkOriginConfiguration)

The proposed Internet and VpcConfiguration to apply to this Amazon S3 access point. VpcConfiguration does not apply to multi-region access points. If the access preview is for a new resource and neither is specified, the access preview uses Internet for the network origin. If the access preview is for an existing resource and neither is specified, the access preview uses the exiting network origin.