Network.AWS.AccessAnalyzer.Types.SecretsManagerSecretConfiguration

Description

Documentation

data SecretsManagerSecretConfiguration

The configuration for a Secrets Manager secret. For more information, see CreateSecret.

You can propose a configuration for a new secret or an existing secret that you own by specifying the secret policy and optional KMS encryption key. If the configuration is for an existing secret and you do not specify the secret policy, the access preview uses the existing policy for the secret. If the access preview is for a new resource and you do not specify the policy, the access preview assumes a secret without a policy. To propose deletion of an existing policy, you can specify an empty string. If the proposed configuration is for a new secret and you do not specify the KMS key ID, the access preview uses the default CMK of the Amazon Web Services account. If you specify an empty string for the KMS key ID, the access preview uses the default CMK of the Amazon Web Services account. For more information about secret policy limits, see Quotas for Secrets Manager..

See: newSecretsManagerSecretConfiguration smart constructor.

Constructors

SecretsManagerSecretConfiguration'

Fields

kmsKeyId :: Maybe Text

The proposed ARN, key ID, or alias of the KMS customer master key (CMK).

secretPolicy :: Maybe Text

The proposed resource policy defining who can access or manage the secret.

newSecretsManagerSecretConfiguration :: SecretsManagerSecretConfiguration

Create a value of SecretsManagerSecretConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:SecretsManagerSecretConfiguration', secretsManagerSecretConfiguration_kmsKeyId - The proposed ARN, key ID, or alias of the KMS customer master key (CMK).

$sel:secretPolicy:SecretsManagerSecretConfiguration', secretsManagerSecretConfiguration_secretPolicy - The proposed resource policy defining who can access or manage the secret.

secretsManagerSecretConfiguration_kmsKeyId :: Lens' SecretsManagerSecretConfiguration (Maybe Text)

The proposed ARN, key ID, or alias of the KMS customer master key (CMK).

secretsManagerSecretConfiguration_secretPolicy :: Lens' SecretsManagerSecretConfiguration (Maybe Text)

The proposed resource policy defining who can access or manage the secret.