Network.AWS.AccessAnalyzer.Types.Trail

Description

Documentation

data Trail

Contains details about the CloudTrail trail being analyzed to generate a policy.

See: newTrail smart constructor.

Constructors

Trail'

Fields

regions :: Maybe [Text]

A list of regions to get CloudTrail data from and analyze to generate a policy.

allRegions :: Maybe Bool

Possible values are true or false. If set to true, IAM Access Analyzer retrieves CloudTrail data from all regions to analyze and generate a policy.

cloudTrailArn :: Text

Specifies the ARN of the trail. The format of a trail ARN is arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail.

newTrail

Arguments

:: Text
$sel:cloudTrailArn:Trail'

-> Trail

Create a value of Trail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regions:Trail', trail_regions - A list of regions to get CloudTrail data from and analyze to generate a policy.

$sel:allRegions:Trail', trail_allRegions - Possible values are true or false. If set to true, IAM Access Analyzer retrieves CloudTrail data from all regions to analyze and generate a policy.

$sel:cloudTrailArn:Trail', trail_cloudTrailArn - Specifies the ARN of the trail. The format of a trail ARN is arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail.

trail_regions :: Lens' Trail (Maybe [Text])

A list of regions to get CloudTrail data from and analyze to generate a policy.

trail_allRegions :: Lens' Trail (Maybe Bool)

Possible values are true or false. If set to true, IAM Access Analyzer retrieves CloudTrail data from all regions to analyze and generate a policy.

trail_cloudTrailArn :: Lens' Trail Text

Specifies the ARN of the trail. The format of a trail ARN is arn:aws:cloudtrail:us-east-2:123456789012:trail/MyTrail.