Network.AWS.AccessAnalyzer.UpdateFindings

Description

Updates the status for the specified findings.

Creating a Request

data UpdateFindings

Updates findings with the new values provided in the request.

See: newUpdateFindings smart constructor.

Constructors

UpdateFindings'

Fields

clientToken :: Maybe Text

A client token.

ids :: Maybe [Text]

The IDs of the findings to update.

resourceArn :: Maybe Text

The ARN of the resource identified in the finding.

analyzerArn :: Text

The ARN of the analyzer that generated the findings to update.

status :: FindingStatusUpdate

The state represents the action to take to update the finding Status. Use ARCHIVE to change an Active finding to an Archived finding. Use ACTIVE to change an Archived finding to an Active finding.

newUpdateFindings

Arguments

:: Text
$sel:analyzerArn:UpdateFindings'

-> FindingStatusUpdate
$sel:status:UpdateFindings'

-> UpdateFindings

Create a value of UpdateFindings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateFindings', updateFindings_clientToken - A client token.

$sel:ids:UpdateFindings', updateFindings_ids - The IDs of the findings to update.

$sel:resourceArn:UpdateFindings', updateFindings_resourceArn - The ARN of the resource identified in the finding.

$sel:analyzerArn:UpdateFindings', updateFindings_analyzerArn - The ARN of the analyzer that generated the findings to update.

$sel:status:UpdateFindings', updateFindings_status - The state represents the action to take to update the finding Status. Use ARCHIVE to change an Active finding to an Archived finding. Use ACTIVE to change an Archived finding to an Active finding.

Request Lenses

updateFindings_clientToken :: Lens' UpdateFindings (Maybe Text)

A client token.

updateFindings_ids :: Lens' UpdateFindings (Maybe [Text])

The IDs of the findings to update.

updateFindings_resourceArn :: Lens' UpdateFindings (Maybe Text)

The ARN of the resource identified in the finding.

updateFindings_analyzerArn :: Lens' UpdateFindings Text

The ARN of the analyzer that generated the findings to update.

updateFindings_status :: Lens' UpdateFindings FindingStatusUpdate

The state represents the action to take to update the finding Status. Use ARCHIVE to change an Active finding to an Archived finding. Use ACTIVE to change an Archived finding to an Active finding.

Destructuring the Response

data UpdateFindingsResponse

See: newUpdateFindingsResponse smart constructor.

Constructors

UpdateFindingsResponse'

newUpdateFindingsResponse :: UpdateFindingsResponse

Create a value of UpdateFindingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.