Network.AWS.Account.GetAlternateContact

Description

Retrieves the specified alternate contact attached to an Amazon Web Services account.

For complete details about how to use the alternate contact operations, see Access or updating the alternate contacts.

Creating a Request

data GetAlternateContact

See: newGetAlternateContact smart constructor.

Constructors

GetAlternateContact'

Fields

accountId :: Maybe Text

Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the organization's management account or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have all features enabled, and the organization must have trusted access enabled for the Account Management service, and optionally a delegated admin account assigned.

The management account can't specify its own AccountId; it must call the operation in standalone context by not including the AccountId parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.

alternateContactType :: AlternateContactType

Specifies which alternate contact you want to retrieve.

newGetAlternateContact

Arguments

:: AlternateContactType
$sel:alternateContactType:GetAlternateContact'

-> GetAlternateContact

Create a value of GetAlternateContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:GetAlternateContact', getAlternateContact_accountId - Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the organization's management account or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have all features enabled, and the organization must have trusted access enabled for the Account Management service, and optionally a delegated admin account assigned.

The management account can't specify its own AccountId; it must call the operation in standalone context by not including the AccountId parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.

$sel:alternateContactType:GetAlternateContact', getAlternateContact_alternateContactType - Specifies which alternate contact you want to retrieve.

Request Lenses

getAlternateContact_accountId :: Lens' GetAlternateContact (Maybe Text)

Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the organization's management account or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have all features enabled, and the organization must have trusted access enabled for the Account Management service, and optionally a delegated admin account assigned.

The management account can't specify its own AccountId; it must call the operation in standalone context by not including the AccountId parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.

getAlternateContact_alternateContactType :: Lens' GetAlternateContact AlternateContactType

Specifies which alternate contact you want to retrieve.

Destructuring the Response

data GetAlternateContactResponse

See: newGetAlternateContactResponse smart constructor.

Constructors

GetAlternateContactResponse'

Fields

alternateContact :: Maybe AlternateContact

A structure that contains the details for the specified alternate contact.

httpStatus :: Int

The response's http status code.

newGetAlternateContactResponse

Arguments

:: Int
$sel:httpStatus:GetAlternateContactResponse'

-> GetAlternateContactResponse

Create a value of GetAlternateContactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alternateContact:GetAlternateContactResponse', getAlternateContactResponse_alternateContact - A structure that contains the details for the specified alternate contact.

$sel:httpStatus:GetAlternateContactResponse', getAlternateContactResponse_httpStatus - The response's http status code.

Response Lenses

getAlternateContactResponse_alternateContact :: Lens' GetAlternateContactResponse (Maybe AlternateContact)

A structure that contains the details for the specified alternate contact.

getAlternateContactResponse_httpStatus :: Lens' GetAlternateContactResponse Int

The response's http status code.