Network.AWS.Account.PutAlternateContact

Description

Modifies the specified alternate contact attached to an Amazon Web Services account.

For complete details about how to use the alternate contact operations, see Access or updating the alternate contacts.

Creating a Request

data PutAlternateContact

See: newPutAlternateContact smart constructor.

Constructors

PutAlternateContact'

Fields

accountId :: Maybe Text

Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the organization's management account or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have all features enabled, and the organization must have trusted access enabled for the Account Management service, and optionally a delegated admin account assigned.

The management account can't specify its own AccountId; it must call the operation in standalone context by not including the AccountId parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.

alternateContactType :: AlternateContactType

Specifies which alternate contact you want to create or update.

emailAddress :: Sensitive Text

Specifies an email address for the alternate contact.

name :: Sensitive Text

Specifies a name for the alternate contact.

phoneNumber :: Sensitive Text

Specifies a phone number for the alternate contact.

title :: Sensitive Text

Specifies a title for the alternate contact.

newPutAlternateContact

Arguments

:: AlternateContactType
$sel:alternateContactType:PutAlternateContact'

-> Text
$sel:emailAddress:PutAlternateContact'

-> Text
$sel:name:PutAlternateContact'

-> Text
$sel:phoneNumber:PutAlternateContact'

-> Text
$sel:title:PutAlternateContact'

-> PutAlternateContact

Create a value of PutAlternateContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:PutAlternateContact', putAlternateContact_accountId - Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the organization's management account or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have all features enabled, and the organization must have trusted access enabled for the Account Management service, and optionally a delegated admin account assigned.

The management account can't specify its own AccountId; it must call the operation in standalone context by not including the AccountId parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.

$sel:alternateContactType:PutAlternateContact', putAlternateContact_alternateContactType - Specifies which alternate contact you want to create or update.

$sel:emailAddress:PutAlternateContact', putAlternateContact_emailAddress - Specifies an email address for the alternate contact.

$sel:name:PutAlternateContact', putAlternateContact_name - Specifies a name for the alternate contact.

$sel:phoneNumber:PutAlternateContact', putAlternateContact_phoneNumber - Specifies a phone number for the alternate contact.

$sel:title:PutAlternateContact', putAlternateContact_title - Specifies a title for the alternate contact.

Request Lenses

putAlternateContact_accountId :: Lens' PutAlternateContact (Maybe Text)

Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the organization's management account or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have all features enabled, and the organization must have trusted access enabled for the Account Management service, and optionally a delegated admin account assigned.

The management account can't specify its own AccountId; it must call the operation in standalone context by not including the AccountId parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.

putAlternateContact_alternateContactType :: Lens' PutAlternateContact AlternateContactType

Specifies which alternate contact you want to create or update.

putAlternateContact_emailAddress :: Lens' PutAlternateContact Text

Specifies an email address for the alternate contact.

putAlternateContact_name :: Lens' PutAlternateContact Text

Specifies a name for the alternate contact.

putAlternateContact_phoneNumber :: Lens' PutAlternateContact Text

Specifies a phone number for the alternate contact.

putAlternateContact_title :: Lens' PutAlternateContact Text

Specifies a title for the alternate contact.

Destructuring the Response

data PutAlternateContactResponse

See: newPutAlternateContactResponse smart constructor.

Constructors

PutAlternateContactResponse'

newPutAlternateContactResponse :: PutAlternateContactResponse

Create a value of PutAlternateContactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.