Network.AWS.AlexaBusiness.AssociateDeviceWithNetworkProfile

Description

Associates a device with the specified network profile.

Creating a Request

data AssociateDeviceWithNetworkProfile

See: newAssociateDeviceWithNetworkProfile smart constructor.

Constructors

AssociateDeviceWithNetworkProfile'

Fields

deviceArn :: Text

The device ARN.

networkProfileArn :: Text

The ARN of the network profile to associate with a device.

newAssociateDeviceWithNetworkProfile

Arguments

:: Text
$sel:deviceArn:AssociateDeviceWithNetworkProfile'

-> Text
$sel:networkProfileArn:AssociateDeviceWithNetworkProfile'

-> AssociateDeviceWithNetworkProfile

Create a value of AssociateDeviceWithNetworkProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceArn:AssociateDeviceWithNetworkProfile', associateDeviceWithNetworkProfile_deviceArn - The device ARN.

$sel:networkProfileArn:AssociateDeviceWithNetworkProfile', associateDeviceWithNetworkProfile_networkProfileArn - The ARN of the network profile to associate with a device.

Request Lenses

associateDeviceWithNetworkProfile_deviceArn :: Lens' AssociateDeviceWithNetworkProfile Text

The device ARN.

associateDeviceWithNetworkProfile_networkProfileArn :: Lens' AssociateDeviceWithNetworkProfile Text

The ARN of the network profile to associate with a device.

Destructuring the Response

data AssociateDeviceWithNetworkProfileResponse

See: newAssociateDeviceWithNetworkProfileResponse smart constructor.

Constructors

AssociateDeviceWithNetworkProfileResponse'

Fields

httpStatus :: Int

The response's http status code.

newAssociateDeviceWithNetworkProfileResponse

Arguments

:: Int
$sel:httpStatus:AssociateDeviceWithNetworkProfileResponse'

-> AssociateDeviceWithNetworkProfileResponse

Create a value of AssociateDeviceWithNetworkProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateDeviceWithNetworkProfileResponse', associateDeviceWithNetworkProfileResponse_httpStatus - The response's http status code.

Response Lenses

associateDeviceWithNetworkProfileResponse_httpStatus :: Lens' AssociateDeviceWithNetworkProfileResponse Int

The response's http status code.