Network.AWS.AlexaBusiness.CreateBusinessReportSchedule

Description

Creates a recurring schedule for usage reports to deliver to the specified S3 location with a specified daily or weekly interval.

Creating a Request

data CreateBusinessReportSchedule

See: newCreateBusinessReportSchedule smart constructor.

Constructors

CreateBusinessReportSchedule'

Fields

s3KeyPrefix :: Maybe Text

The S3 key where the report is delivered.

recurrence :: Maybe BusinessReportRecurrence

The recurrence of the reports. If this isn't specified, the report will only be delivered one time when the API is called.

tags :: Maybe [Tag]

The tags for the business report schedule.

s3BucketName :: Maybe Text

The S3 bucket name of the output reports. If this isn't specified, the report can be retrieved from a download link by calling ListBusinessReportSchedule.

clientRequestToken :: Maybe Text

The client request token.

scheduleName :: Maybe Text

The name identifier of the schedule.

format :: BusinessReportFormat

The format of the generated report (individual CSV files or zipped files of individual files).

contentRange :: BusinessReportContentRange

The content range of the reports.

newCreateBusinessReportSchedule

Arguments

:: BusinessReportFormat
$sel:format:CreateBusinessReportSchedule'

-> BusinessReportContentRange
$sel:contentRange:CreateBusinessReportSchedule'

-> CreateBusinessReportSchedule

Create a value of CreateBusinessReportSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3KeyPrefix:CreateBusinessReportSchedule', createBusinessReportSchedule_s3KeyPrefix - The S3 key where the report is delivered.

$sel:recurrence:CreateBusinessReportSchedule', createBusinessReportSchedule_recurrence - The recurrence of the reports. If this isn't specified, the report will only be delivered one time when the API is called.

$sel:tags:CreateBusinessReportSchedule', createBusinessReportSchedule_tags - The tags for the business report schedule.

$sel:s3BucketName:CreateBusinessReportSchedule', createBusinessReportSchedule_s3BucketName - The S3 bucket name of the output reports. If this isn't specified, the report can be retrieved from a download link by calling ListBusinessReportSchedule.

$sel:clientRequestToken:CreateBusinessReportSchedule', createBusinessReportSchedule_clientRequestToken - The client request token.

$sel:scheduleName:CreateBusinessReportSchedule', createBusinessReportSchedule_scheduleName - The name identifier of the schedule.

$sel:format:CreateBusinessReportSchedule', createBusinessReportSchedule_format - The format of the generated report (individual CSV files or zipped files of individual files).

$sel:contentRange:CreateBusinessReportSchedule', createBusinessReportSchedule_contentRange - The content range of the reports.

Request Lenses

createBusinessReportSchedule_s3KeyPrefix :: Lens' CreateBusinessReportSchedule (Maybe Text)

The S3 key where the report is delivered.

createBusinessReportSchedule_recurrence :: Lens' CreateBusinessReportSchedule (Maybe BusinessReportRecurrence)

The recurrence of the reports. If this isn't specified, the report will only be delivered one time when the API is called.

createBusinessReportSchedule_tags :: Lens' CreateBusinessReportSchedule (Maybe [Tag])

The tags for the business report schedule.

createBusinessReportSchedule_s3BucketName :: Lens' CreateBusinessReportSchedule (Maybe Text)

The S3 bucket name of the output reports. If this isn't specified, the report can be retrieved from a download link by calling ListBusinessReportSchedule.

createBusinessReportSchedule_clientRequestToken :: Lens' CreateBusinessReportSchedule (Maybe Text)

The client request token.

createBusinessReportSchedule_scheduleName :: Lens' CreateBusinessReportSchedule (Maybe Text)

The name identifier of the schedule.

createBusinessReportSchedule_format :: Lens' CreateBusinessReportSchedule BusinessReportFormat

The format of the generated report (individual CSV files or zipped files of individual files).

createBusinessReportSchedule_contentRange :: Lens' CreateBusinessReportSchedule BusinessReportContentRange

The content range of the reports.

Destructuring the Response

data CreateBusinessReportScheduleResponse

See: newCreateBusinessReportScheduleResponse smart constructor.

Constructors

CreateBusinessReportScheduleResponse'

Fields

scheduleArn :: Maybe Text

The ARN of the business report schedule.

httpStatus :: Int

The response's http status code.

newCreateBusinessReportScheduleResponse

Arguments

:: Int
$sel:httpStatus:CreateBusinessReportScheduleResponse'

-> CreateBusinessReportScheduleResponse

Create a value of CreateBusinessReportScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduleArn:CreateBusinessReportScheduleResponse', createBusinessReportScheduleResponse_scheduleArn - The ARN of the business report schedule.

$sel:httpStatus:CreateBusinessReportScheduleResponse', createBusinessReportScheduleResponse_httpStatus - The response's http status code.

Response Lenses

createBusinessReportScheduleResponse_scheduleArn :: Lens' CreateBusinessReportScheduleResponse (Maybe Text)

The ARN of the business report schedule.

createBusinessReportScheduleResponse_httpStatus :: Lens' CreateBusinessReportScheduleResponse Int

The response's http status code.