Network.AWS.AlexaBusiness.CreateConferenceProvider

Description

Adds a new conference provider under the user's AWS account.

Creating a Request

data CreateConferenceProvider

See: newCreateConferenceProvider smart constructor.

Constructors

CreateConferenceProvider'

Fields

pSTNDialIn :: Maybe PSTNDialIn

The information for PSTN conferencing.

clientRequestToken :: Maybe Text

The request token of the client.

iPDialIn :: Maybe IPDialIn

The IP endpoint and protocol for calling.

tags :: Maybe [Tag]

The tags to be added to the specified resource. Do not provide system tags.

conferenceProviderName :: Text

The name of the conference provider.

conferenceProviderType :: ConferenceProviderType

Represents a type within a list of predefined types.

meetingSetting :: MeetingSetting

The meeting settings for the conference provider.

newCreateConferenceProvider

Arguments

:: Text
$sel:conferenceProviderName:CreateConferenceProvider'

-> ConferenceProviderType
$sel:conferenceProviderType:CreateConferenceProvider'

-> MeetingSetting
$sel:meetingSetting:CreateConferenceProvider'

-> CreateConferenceProvider

Create a value of CreateConferenceProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pSTNDialIn:CreateConferenceProvider', createConferenceProvider_pSTNDialIn - The information for PSTN conferencing.

$sel:clientRequestToken:CreateConferenceProvider', createConferenceProvider_clientRequestToken - The request token of the client.

$sel:iPDialIn:CreateConferenceProvider', createConferenceProvider_iPDialIn - The IP endpoint and protocol for calling.

$sel:tags:CreateConferenceProvider', createConferenceProvider_tags - The tags to be added to the specified resource. Do not provide system tags.

$sel:conferenceProviderName:CreateConferenceProvider', createConferenceProvider_conferenceProviderName - The name of the conference provider.

$sel:conferenceProviderType:CreateConferenceProvider', createConferenceProvider_conferenceProviderType - Represents a type within a list of predefined types.

$sel:meetingSetting:CreateConferenceProvider', createConferenceProvider_meetingSetting - The meeting settings for the conference provider.

Request Lenses

createConferenceProvider_pSTNDialIn :: Lens' CreateConferenceProvider (Maybe PSTNDialIn)

The information for PSTN conferencing.

createConferenceProvider_clientRequestToken :: Lens' CreateConferenceProvider (Maybe Text)

The request token of the client.

createConferenceProvider_iPDialIn :: Lens' CreateConferenceProvider (Maybe IPDialIn)

The IP endpoint and protocol for calling.

createConferenceProvider_tags :: Lens' CreateConferenceProvider (Maybe [Tag])

The tags to be added to the specified resource. Do not provide system tags.

createConferenceProvider_conferenceProviderName :: Lens' CreateConferenceProvider Text

The name of the conference provider.

createConferenceProvider_conferenceProviderType :: Lens' CreateConferenceProvider ConferenceProviderType

Represents a type within a list of predefined types.

createConferenceProvider_meetingSetting :: Lens' CreateConferenceProvider MeetingSetting

The meeting settings for the conference provider.

Destructuring the Response

data CreateConferenceProviderResponse

See: newCreateConferenceProviderResponse smart constructor.

Constructors

CreateConferenceProviderResponse'

Fields

conferenceProviderArn :: Maybe Text

The ARN of the newly-created conference provider.

httpStatus :: Int

The response's http status code.

newCreateConferenceProviderResponse

Arguments

:: Int
$sel:httpStatus:CreateConferenceProviderResponse'

-> CreateConferenceProviderResponse

Create a value of CreateConferenceProviderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conferenceProviderArn:CreateConferenceProviderResponse', createConferenceProviderResponse_conferenceProviderArn - The ARN of the newly-created conference provider.

$sel:httpStatus:CreateConferenceProviderResponse', createConferenceProviderResponse_httpStatus - The response's http status code.

Response Lenses

createConferenceProviderResponse_conferenceProviderArn :: Lens' CreateConferenceProviderResponse (Maybe Text)

The ARN of the newly-created conference provider.

createConferenceProviderResponse_httpStatus :: Lens' CreateConferenceProviderResponse Int

The response's http status code.