Network.AWS.AlexaBusiness.CreateNetworkProfile

Description

Creates a network profile with the specified details.

Creating a Request

data CreateNetworkProfile

See: newCreateNetworkProfile smart constructor.

Constructors

CreateNetworkProfile'

Fields

certificateAuthorityArn :: Maybe Text

The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

trustAnchors :: Maybe (NonEmpty Text)

The root certificates of your authentication server that is installed on your devices and used to trust your authentication server during EAP negotiation.

currentPassword :: Maybe (Sensitive Text)

The current password of the Wi-Fi network.

eapMethod :: Maybe NetworkEapMethod

The authentication standard that is used in the EAP framework. Currently, EAP_TLS is supported.

tags :: Maybe [Tag]

The tags to be added to the specified resource. Do not provide system tags.

description :: Maybe Text

Detailed information about a device's network profile.

nextPassword :: Maybe (Sensitive Text)

The next, or subsequent, password of the Wi-Fi network. This password is asynchronously transmitted to the device and is used when the password of the network changes to NextPassword.

networkProfileName :: Text

The name of the network profile associated with a device.

ssid :: Text

The SSID of the Wi-Fi network.

securityType :: NetworkSecurityType

The security type of the Wi-Fi network. This can be WPA2_ENTERPRISE, WPA2_PSK, WPA_PSK, WEP, or OPEN.

clientRequestToken :: Text

newCreateNetworkProfile

Arguments

:: Text
$sel:networkProfileName:CreateNetworkProfile'

-> Text
$sel:ssid:CreateNetworkProfile'

-> NetworkSecurityType
$sel:securityType:CreateNetworkProfile'

-> Text
$sel:clientRequestToken:CreateNetworkProfile'

-> CreateNetworkProfile

Create a value of CreateNetworkProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:CreateNetworkProfile', createNetworkProfile_certificateAuthorityArn - The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

$sel:trustAnchors:CreateNetworkProfile', createNetworkProfile_trustAnchors - The root certificates of your authentication server that is installed on your devices and used to trust your authentication server during EAP negotiation.

$sel:currentPassword:CreateNetworkProfile', createNetworkProfile_currentPassword - The current password of the Wi-Fi network.

$sel:eapMethod:CreateNetworkProfile', createNetworkProfile_eapMethod - The authentication standard that is used in the EAP framework. Currently, EAP_TLS is supported.

$sel:tags:CreateNetworkProfile', createNetworkProfile_tags - The tags to be added to the specified resource. Do not provide system tags.

$sel:description:CreateNetworkProfile', createNetworkProfile_description - Detailed information about a device's network profile.

$sel:nextPassword:CreateNetworkProfile', createNetworkProfile_nextPassword - The next, or subsequent, password of the Wi-Fi network. This password is asynchronously transmitted to the device and is used when the password of the network changes to NextPassword.

$sel:networkProfileName:CreateNetworkProfile', createNetworkProfile_networkProfileName - The name of the network profile associated with a device.

$sel:ssid:CreateNetworkProfile', createNetworkProfile_ssid - The SSID of the Wi-Fi network.

$sel:securityType:CreateNetworkProfile', createNetworkProfile_securityType - The security type of the Wi-Fi network. This can be WPA2_ENTERPRISE, WPA2_PSK, WPA_PSK, WEP, or OPEN.

$sel:clientRequestToken:CreateNetworkProfile', createNetworkProfile_clientRequestToken - Undocumented member.

Request Lenses

createNetworkProfile_certificateAuthorityArn :: Lens' CreateNetworkProfile (Maybe Text)

The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

createNetworkProfile_trustAnchors :: Lens' CreateNetworkProfile (Maybe (NonEmpty Text))

The root certificates of your authentication server that is installed on your devices and used to trust your authentication server during EAP negotiation.

createNetworkProfile_currentPassword :: Lens' CreateNetworkProfile (Maybe Text)

The current password of the Wi-Fi network.

createNetworkProfile_eapMethod :: Lens' CreateNetworkProfile (Maybe NetworkEapMethod)

The authentication standard that is used in the EAP framework. Currently, EAP_TLS is supported.

createNetworkProfile_tags :: Lens' CreateNetworkProfile (Maybe [Tag])

The tags to be added to the specified resource. Do not provide system tags.

createNetworkProfile_description :: Lens' CreateNetworkProfile (Maybe Text)

Detailed information about a device's network profile.

createNetworkProfile_nextPassword :: Lens' CreateNetworkProfile (Maybe Text)

The next, or subsequent, password of the Wi-Fi network. This password is asynchronously transmitted to the device and is used when the password of the network changes to NextPassword.

createNetworkProfile_networkProfileName :: Lens' CreateNetworkProfile Text

The name of the network profile associated with a device.

createNetworkProfile_ssid :: Lens' CreateNetworkProfile Text

The SSID of the Wi-Fi network.

createNetworkProfile_securityType :: Lens' CreateNetworkProfile NetworkSecurityType

The security type of the Wi-Fi network. This can be WPA2_ENTERPRISE, WPA2_PSK, WPA_PSK, WEP, or OPEN.

createNetworkProfile_clientRequestToken :: Lens' CreateNetworkProfile Text

Undocumented member.

Destructuring the Response

data CreateNetworkProfileResponse

See: newCreateNetworkProfileResponse smart constructor.

Constructors

CreateNetworkProfileResponse'

Fields

networkProfileArn :: Maybe Text

The ARN of the network profile associated with a device.

httpStatus :: Int

The response's http status code.

newCreateNetworkProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateNetworkProfileResponse'

-> CreateNetworkProfileResponse

Create a value of CreateNetworkProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkProfileArn:CreateNetworkProfileResponse', createNetworkProfileResponse_networkProfileArn - The ARN of the network profile associated with a device.

$sel:httpStatus:CreateNetworkProfileResponse', createNetworkProfileResponse_httpStatus - The response's http status code.

Response Lenses

createNetworkProfileResponse_networkProfileArn :: Lens' CreateNetworkProfileResponse (Maybe Text)

The ARN of the network profile associated with a device.

createNetworkProfileResponse_httpStatus :: Lens' CreateNetworkProfileResponse Int

The response's http status code.