Network.AWS.AlexaBusiness.CreateProfile

Description

Creates a new room profile with the specified details.

Creating a Request

data CreateProfile

See: newCreateProfile smart constructor.

Constructors

CreateProfile'

Fields

locale :: Maybe Text

The locale of the room profile. (This is currently only available to a limited preview audience.)

pSTNEnabled :: Maybe Bool

Whether PSTN calling is enabled.

setupModeDisabled :: Maybe Bool

Whether room profile setup is enabled.

maxVolumeLimit :: Maybe Int

The maximum volume limit for a room profile.

tags :: Maybe [Tag]

The tags for the profile.

meetingRoomConfiguration :: Maybe CreateMeetingRoomConfiguration

The meeting room settings of a room profile.

clientRequestToken :: Maybe Text

The user-specified token that is used during the creation of a profile.

dataRetentionOptIn :: Maybe Bool

Whether data retention of the profile is enabled.

profileName :: Text

The name of a room profile.

timezone :: Text

The time zone used by a room profile.

address :: Text

The valid address for the room.

distanceUnit :: DistanceUnit

The distance unit to be used by devices in the profile.

temperatureUnit :: TemperatureUnit

The temperature unit to be used by devices in the profile.

wakeWord :: WakeWord

A wake word for Alexa, Echo, Amazon, or a computer.

newCreateProfile

Arguments

:: Text
$sel:profileName:CreateProfile'

-> Text
$sel:timezone:CreateProfile'

-> Text
$sel:address:CreateProfile'

-> DistanceUnit
$sel:distanceUnit:CreateProfile'

-> TemperatureUnit
$sel:temperatureUnit:CreateProfile'

-> WakeWord
$sel:wakeWord:CreateProfile'

-> CreateProfile

Create a value of CreateProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locale:CreateProfile', createProfile_locale - The locale of the room profile. (This is currently only available to a limited preview audience.)

$sel:pSTNEnabled:CreateProfile', createProfile_pSTNEnabled - Whether PSTN calling is enabled.

$sel:setupModeDisabled:CreateProfile', createProfile_setupModeDisabled - Whether room profile setup is enabled.

$sel:maxVolumeLimit:CreateProfile', createProfile_maxVolumeLimit - The maximum volume limit for a room profile.

$sel:tags:CreateProfile', createProfile_tags - The tags for the profile.

$sel:meetingRoomConfiguration:CreateProfile', createProfile_meetingRoomConfiguration - The meeting room settings of a room profile.

$sel:clientRequestToken:CreateProfile', createProfile_clientRequestToken - The user-specified token that is used during the creation of a profile.

$sel:dataRetentionOptIn:CreateProfile', createProfile_dataRetentionOptIn - Whether data retention of the profile is enabled.

$sel:profileName:CreateProfile', createProfile_profileName - The name of a room profile.

$sel:timezone:CreateProfile', createProfile_timezone - The time zone used by a room profile.

$sel:address:CreateProfile', createProfile_address - The valid address for the room.

$sel:distanceUnit:CreateProfile', createProfile_distanceUnit - The distance unit to be used by devices in the profile.

$sel:temperatureUnit:CreateProfile', createProfile_temperatureUnit - The temperature unit to be used by devices in the profile.

$sel:wakeWord:CreateProfile', createProfile_wakeWord - A wake word for Alexa, Echo, Amazon, or a computer.

Request Lenses

createProfile_locale :: Lens' CreateProfile (Maybe Text)

The locale of the room profile. (This is currently only available to a limited preview audience.)

createProfile_pSTNEnabled :: Lens' CreateProfile (Maybe Bool)

Whether PSTN calling is enabled.

createProfile_setupModeDisabled :: Lens' CreateProfile (Maybe Bool)

Whether room profile setup is enabled.

createProfile_maxVolumeLimit :: Lens' CreateProfile (Maybe Int)

The maximum volume limit for a room profile.

createProfile_tags :: Lens' CreateProfile (Maybe [Tag])

The tags for the profile.

createProfile_meetingRoomConfiguration :: Lens' CreateProfile (Maybe CreateMeetingRoomConfiguration)

The meeting room settings of a room profile.

createProfile_clientRequestToken :: Lens' CreateProfile (Maybe Text)

The user-specified token that is used during the creation of a profile.

createProfile_dataRetentionOptIn :: Lens' CreateProfile (Maybe Bool)

Whether data retention of the profile is enabled.

createProfile_profileName :: Lens' CreateProfile Text

The name of a room profile.

createProfile_timezone :: Lens' CreateProfile Text

The time zone used by a room profile.

createProfile_address :: Lens' CreateProfile Text

The valid address for the room.

createProfile_distanceUnit :: Lens' CreateProfile DistanceUnit

The distance unit to be used by devices in the profile.

createProfile_temperatureUnit :: Lens' CreateProfile TemperatureUnit

The temperature unit to be used by devices in the profile.

createProfile_wakeWord :: Lens' CreateProfile WakeWord

A wake word for Alexa, Echo, Amazon, or a computer.

Destructuring the Response

data CreateProfileResponse

See: newCreateProfileResponse smart constructor.

Constructors

CreateProfileResponse'

Fields

profileArn :: Maybe Text

The ARN of the newly created room profile in the response.

httpStatus :: Int

The response's http status code.

newCreateProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateProfileResponse'

-> CreateProfileResponse

Create a value of CreateProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profileArn:CreateProfileResponse', createProfileResponse_profileArn - The ARN of the newly created room profile in the response.

$sel:httpStatus:CreateProfileResponse', createProfileResponse_httpStatus - The response's http status code.

Response Lenses

createProfileResponse_profileArn :: Lens' CreateProfileResponse (Maybe Text)

The ARN of the newly created room profile in the response.

createProfileResponse_httpStatus :: Lens' CreateProfileResponse Int

The response's http status code.