Network.AWS.AlexaBusiness.CreateRoom

Description

Creates a room with the specified details.

Creating a Request

data CreateRoom

See: newCreateRoom smart constructor.

Constructors

CreateRoom'

Fields

tags :: Maybe [Tag]

The tags for the room.

profileArn :: Maybe Text

The profile ARN for the room. This is required.

description :: Maybe Text

The description for the room.

providerCalendarId :: Maybe Text

The calendar ARN for the room.

clientRequestToken :: Maybe Text

A unique, user-specified identifier for this request that ensures idempotency.

roomName :: Text

The name for the room.

newCreateRoom

Arguments

:: Text
$sel:roomName:CreateRoom'

-> CreateRoom

Create a value of CreateRoom with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateRoom', createRoom_tags - The tags for the room.

$sel:profileArn:CreateRoom', createRoom_profileArn - The profile ARN for the room. This is required.

$sel:description:CreateRoom', createRoom_description - The description for the room.

$sel:providerCalendarId:CreateRoom', createRoom_providerCalendarId - The calendar ARN for the room.

$sel:clientRequestToken:CreateRoom', createRoom_clientRequestToken - A unique, user-specified identifier for this request that ensures idempotency.

$sel:roomName:CreateRoom', createRoom_roomName - The name for the room.

Request Lenses

createRoom_tags :: Lens' CreateRoom (Maybe [Tag])

The tags for the room.

createRoom_profileArn :: Lens' CreateRoom (Maybe Text)

The profile ARN for the room. This is required.

createRoom_description :: Lens' CreateRoom (Maybe Text)

The description for the room.

createRoom_providerCalendarId :: Lens' CreateRoom (Maybe Text)

The calendar ARN for the room.

createRoom_clientRequestToken :: Lens' CreateRoom (Maybe Text)

A unique, user-specified identifier for this request that ensures idempotency.

createRoom_roomName :: Lens' CreateRoom Text

The name for the room.

Destructuring the Response

data CreateRoomResponse

See: newCreateRoomResponse smart constructor.

Constructors

CreateRoomResponse'

Fields

roomArn :: Maybe Text

The ARN of the newly created room in the response.

httpStatus :: Int

The response's http status code.

newCreateRoomResponse

Arguments

:: Int
$sel:httpStatus:CreateRoomResponse'

-> CreateRoomResponse

Create a value of CreateRoomResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roomArn:CreateRoomResponse', createRoomResponse_roomArn - The ARN of the newly created room in the response.

$sel:httpStatus:CreateRoomResponse', createRoomResponse_httpStatus - The response's http status code.

Response Lenses

createRoomResponse_roomArn :: Lens' CreateRoomResponse (Maybe Text)

The ARN of the newly created room in the response.

createRoomResponse_httpStatus :: Lens' CreateRoomResponse Int

The response's http status code.