Network.AWS.AlexaBusiness.CreateSkillGroup

Description

Creates a skill group with a specified name and description.

Creating a Request

data CreateSkillGroup

See: newCreateSkillGroup smart constructor.

Constructors

CreateSkillGroup'

Fields

tags :: Maybe [Tag]

The tags for the skill group.

description :: Maybe Text

The description for the skill group.

clientRequestToken :: Maybe Text

A unique, user-specified identifier for this request that ensures idempotency.

skillGroupName :: Text

The name for the skill group.

newCreateSkillGroup

Arguments

:: Text
$sel:skillGroupName:CreateSkillGroup'

-> CreateSkillGroup

Create a value of CreateSkillGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateSkillGroup', createSkillGroup_tags - The tags for the skill group.

$sel:description:CreateSkillGroup', createSkillGroup_description - The description for the skill group.

$sel:clientRequestToken:CreateSkillGroup', createSkillGroup_clientRequestToken - A unique, user-specified identifier for this request that ensures idempotency.

$sel:skillGroupName:CreateSkillGroup', createSkillGroup_skillGroupName - The name for the skill group.

Request Lenses

createSkillGroup_tags :: Lens' CreateSkillGroup (Maybe [Tag])

The tags for the skill group.

createSkillGroup_description :: Lens' CreateSkillGroup (Maybe Text)

The description for the skill group.

createSkillGroup_clientRequestToken :: Lens' CreateSkillGroup (Maybe Text)

A unique, user-specified identifier for this request that ensures idempotency.

createSkillGroup_skillGroupName :: Lens' CreateSkillGroup Text

The name for the skill group.

Destructuring the Response

data CreateSkillGroupResponse

See: newCreateSkillGroupResponse smart constructor.

Constructors

CreateSkillGroupResponse'

Fields

skillGroupArn :: Maybe Text

The ARN of the newly created skill group in the response.

httpStatus :: Int

The response's http status code.

newCreateSkillGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateSkillGroupResponse'

-> CreateSkillGroupResponse

Create a value of CreateSkillGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:skillGroupArn:CreateSkillGroupResponse', createSkillGroupResponse_skillGroupArn - The ARN of the newly created skill group in the response.

$sel:httpStatus:CreateSkillGroupResponse', createSkillGroupResponse_httpStatus - The response's http status code.

Response Lenses

createSkillGroupResponse_skillGroupArn :: Lens' CreateSkillGroupResponse (Maybe Text)

The ARN of the newly created skill group in the response.

createSkillGroupResponse_httpStatus :: Lens' CreateSkillGroupResponse Int

The response's http status code.