Network.AWS.AlexaBusiness.DeleteBusinessReportSchedule

Description

Deletes the recurring report delivery schedule with the specified schedule ARN.

Creating a Request

data DeleteBusinessReportSchedule

See: newDeleteBusinessReportSchedule smart constructor.

Constructors

DeleteBusinessReportSchedule'

Fields

scheduleArn :: Text

The ARN of the business report schedule.

newDeleteBusinessReportSchedule

Arguments

:: Text
$sel:scheduleArn:DeleteBusinessReportSchedule'

-> DeleteBusinessReportSchedule

Create a value of DeleteBusinessReportSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduleArn:DeleteBusinessReportSchedule', deleteBusinessReportSchedule_scheduleArn - The ARN of the business report schedule.

Request Lenses

deleteBusinessReportSchedule_scheduleArn :: Lens' DeleteBusinessReportSchedule Text

The ARN of the business report schedule.

Destructuring the Response

data DeleteBusinessReportScheduleResponse

See: newDeleteBusinessReportScheduleResponse smart constructor.

Constructors

DeleteBusinessReportScheduleResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteBusinessReportScheduleResponse

Arguments

:: Int
$sel:httpStatus:DeleteBusinessReportScheduleResponse'

-> DeleteBusinessReportScheduleResponse

Create a value of DeleteBusinessReportScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteBusinessReportScheduleResponse', deleteBusinessReportScheduleResponse_httpStatus - The response's http status code.

Response Lenses

deleteBusinessReportScheduleResponse_httpStatus :: Lens' DeleteBusinessReportScheduleResponse Int

The response's http status code.