Network.AWS.AlexaBusiness.ListBusinessReportSchedules

Description

Lists the details of the schedules that a user configured. A download URL of the report associated with each schedule is returned every time this action is called. A new download URL is returned each time, and is valid for 24 hours.

This operation returns paginated results.

Creating a Request

data ListBusinessReportSchedules

See: newListBusinessReportSchedules smart constructor.

Constructors

ListBusinessReportSchedules'

Fields

nextToken :: Maybe Text

The token used to list the remaining schedules from the previous API call.

maxResults :: Maybe Natural

The maximum number of schedules listed in the call.

newListBusinessReportSchedules :: ListBusinessReportSchedules

Create a value of ListBusinessReportSchedules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBusinessReportSchedules', listBusinessReportSchedules_nextToken - The token used to list the remaining schedules from the previous API call.

$sel:maxResults:ListBusinessReportSchedules', listBusinessReportSchedules_maxResults - The maximum number of schedules listed in the call.

Request Lenses

listBusinessReportSchedules_nextToken :: Lens' ListBusinessReportSchedules (Maybe Text)

The token used to list the remaining schedules from the previous API call.

listBusinessReportSchedules_maxResults :: Lens' ListBusinessReportSchedules (Maybe Natural)

The maximum number of schedules listed in the call.

Destructuring the Response

data ListBusinessReportSchedulesResponse

See: newListBusinessReportSchedulesResponse smart constructor.

Constructors

ListBusinessReportSchedulesResponse'

Fields

businessReportSchedules :: Maybe [BusinessReportSchedule]

The schedule of the reports.

nextToken :: Maybe Text

The token used to list the remaining schedules from the previous API call.

httpStatus :: Int

The response's http status code.

newListBusinessReportSchedulesResponse

Arguments

:: Int
$sel:httpStatus:ListBusinessReportSchedulesResponse'

-> ListBusinessReportSchedulesResponse

Create a value of ListBusinessReportSchedulesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:businessReportSchedules:ListBusinessReportSchedulesResponse', listBusinessReportSchedulesResponse_businessReportSchedules - The schedule of the reports.

$sel:nextToken:ListBusinessReportSchedules', listBusinessReportSchedulesResponse_nextToken - The token used to list the remaining schedules from the previous API call.

$sel:httpStatus:ListBusinessReportSchedulesResponse', listBusinessReportSchedulesResponse_httpStatus - The response's http status code.

Response Lenses

listBusinessReportSchedulesResponse_businessReportSchedules :: Lens' ListBusinessReportSchedulesResponse (Maybe [BusinessReportSchedule])

The schedule of the reports.

listBusinessReportSchedulesResponse_nextToken :: Lens' ListBusinessReportSchedulesResponse (Maybe Text)

The token used to list the remaining schedules from the previous API call.

listBusinessReportSchedulesResponse_httpStatus :: Lens' ListBusinessReportSchedulesResponse Int

The response's http status code.