Network.AWS.AlexaBusiness.ListConferenceProviders

Description

Lists conference providers under a specific AWS account.

This operation returns paginated results.

Creating a Request

data ListConferenceProviders

See: newListConferenceProviders smart constructor.

Constructors

ListConferenceProviders'

Fields

nextToken :: Maybe Text

The tokens used for pagination.

maxResults :: Maybe Natural

The maximum number of conference providers to be returned, per paginated calls.

newListConferenceProviders :: ListConferenceProviders

Create a value of ListConferenceProviders with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListConferenceProviders', listConferenceProviders_nextToken - The tokens used for pagination.

$sel:maxResults:ListConferenceProviders', listConferenceProviders_maxResults - The maximum number of conference providers to be returned, per paginated calls.

Request Lenses

listConferenceProviders_nextToken :: Lens' ListConferenceProviders (Maybe Text)

The tokens used for pagination.

listConferenceProviders_maxResults :: Lens' ListConferenceProviders (Maybe Natural)

The maximum number of conference providers to be returned, per paginated calls.

Destructuring the Response

data ListConferenceProvidersResponse

See: newListConferenceProvidersResponse smart constructor.

Constructors

ListConferenceProvidersResponse'

Fields

nextToken :: Maybe Text

The tokens used for pagination.

conferenceProviders :: Maybe [ConferenceProvider]

The conference providers.

httpStatus :: Int

The response's http status code.

newListConferenceProvidersResponse

Arguments

:: Int
$sel:httpStatus:ListConferenceProvidersResponse'

-> ListConferenceProvidersResponse

Create a value of ListConferenceProvidersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListConferenceProviders', listConferenceProvidersResponse_nextToken - The tokens used for pagination.

$sel:conferenceProviders:ListConferenceProvidersResponse', listConferenceProvidersResponse_conferenceProviders - The conference providers.

$sel:httpStatus:ListConferenceProvidersResponse', listConferenceProvidersResponse_httpStatus - The response's http status code.

Response Lenses

listConferenceProvidersResponse_nextToken :: Lens' ListConferenceProvidersResponse (Maybe Text)

The tokens used for pagination.

listConferenceProvidersResponse_conferenceProviders :: Lens' ListConferenceProvidersResponse (Maybe [ConferenceProvider])

The conference providers.

listConferenceProvidersResponse_httpStatus :: Lens' ListConferenceProvidersResponse Int

The response's http status code.