Network.AWS.AlexaBusiness.ListGatewayGroups

Description

Retrieves a list of gateway group summaries. Use GetGatewayGroup to retrieve details of a specific gateway group.

Creating a Request

data ListGatewayGroups

See: newListGatewayGroups smart constructor.

Constructors

ListGatewayGroups'

Fields

nextToken :: Maybe Text

The token used to paginate though multiple pages of gateway group summaries.

maxResults :: Maybe Natural

The maximum number of gateway group summaries to return. The default is 50.

newListGatewayGroups :: ListGatewayGroups

Create a value of ListGatewayGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGatewayGroups', listGatewayGroups_nextToken - The token used to paginate though multiple pages of gateway group summaries.

$sel:maxResults:ListGatewayGroups', listGatewayGroups_maxResults - The maximum number of gateway group summaries to return. The default is 50.

Request Lenses

listGatewayGroups_nextToken :: Lens' ListGatewayGroups (Maybe Text)

The token used to paginate though multiple pages of gateway group summaries.

listGatewayGroups_maxResults :: Lens' ListGatewayGroups (Maybe Natural)

The maximum number of gateway group summaries to return. The default is 50.

Destructuring the Response

data ListGatewayGroupsResponse

See: newListGatewayGroupsResponse smart constructor.

Constructors

ListGatewayGroupsResponse'

Fields

gatewayGroups :: Maybe [GatewayGroupSummary]

The gateway groups in the list.

nextToken :: Maybe Text

The token used to paginate though multiple pages of gateway group summaries.

httpStatus :: Int

The response's http status code.

newListGatewayGroupsResponse

Arguments

:: Int
$sel:httpStatus:ListGatewayGroupsResponse'

-> ListGatewayGroupsResponse

Create a value of ListGatewayGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayGroups:ListGatewayGroupsResponse', listGatewayGroupsResponse_gatewayGroups - The gateway groups in the list.

$sel:nextToken:ListGatewayGroups', listGatewayGroupsResponse_nextToken - The token used to paginate though multiple pages of gateway group summaries.

$sel:httpStatus:ListGatewayGroupsResponse', listGatewayGroupsResponse_httpStatus - The response's http status code.

Response Lenses

listGatewayGroupsResponse_gatewayGroups :: Lens' ListGatewayGroupsResponse (Maybe [GatewayGroupSummary])

The gateway groups in the list.

listGatewayGroupsResponse_nextToken :: Lens' ListGatewayGroupsResponse (Maybe Text)

The token used to paginate though multiple pages of gateway group summaries.

listGatewayGroupsResponse_httpStatus :: Lens' ListGatewayGroupsResponse Int

The response's http status code.