Network.AWS.AlexaBusiness.ListGateways

Description

Retrieves a list of gateway summaries. Use GetGateway to retrieve details of a specific gateway. An optional gateway group ARN can be provided to only retrieve gateway summaries of gateways that are associated with that gateway group ARN.

Creating a Request

data ListGateways

See: newListGateways smart constructor.

Constructors

ListGateways'

Fields

nextToken :: Maybe Text

The token used to paginate though multiple pages of gateway summaries.

maxResults :: Maybe Natural

The maximum number of gateway summaries to return. The default is 50.

gatewayGroupArn :: Maybe Text

The gateway group ARN for which to list gateways.

newListGateways :: ListGateways

Create a value of ListGateways with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGateways', listGateways_nextToken - The token used to paginate though multiple pages of gateway summaries.

$sel:maxResults:ListGateways', listGateways_maxResults - The maximum number of gateway summaries to return. The default is 50.

$sel:gatewayGroupArn:ListGateways', listGateways_gatewayGroupArn - The gateway group ARN for which to list gateways.

Request Lenses

listGateways_nextToken :: Lens' ListGateways (Maybe Text)

The token used to paginate though multiple pages of gateway summaries.

listGateways_maxResults :: Lens' ListGateways (Maybe Natural)

The maximum number of gateway summaries to return. The default is 50.

listGateways_gatewayGroupArn :: Lens' ListGateways (Maybe Text)

The gateway group ARN for which to list gateways.

Destructuring the Response

data ListGatewaysResponse

See: newListGatewaysResponse smart constructor.

Constructors

ListGatewaysResponse'

Fields

nextToken :: Maybe Text

The token used to paginate though multiple pages of gateway summaries.

gateways :: Maybe [GatewaySummary]

The gateways in the list.

httpStatus :: Int

The response's http status code.

newListGatewaysResponse

Arguments

:: Int
$sel:httpStatus:ListGatewaysResponse'

-> ListGatewaysResponse

Create a value of ListGatewaysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGateways', listGatewaysResponse_nextToken - The token used to paginate though multiple pages of gateway summaries.

$sel:gateways:ListGatewaysResponse', listGatewaysResponse_gateways - The gateways in the list.

$sel:httpStatus:ListGatewaysResponse', listGatewaysResponse_httpStatus - The response's http status code.

Response Lenses

listGatewaysResponse_nextToken :: Lens' ListGatewaysResponse (Maybe Text)

The token used to paginate though multiple pages of gateway summaries.

listGatewaysResponse_gateways :: Lens' ListGatewaysResponse (Maybe [GatewaySummary])

The gateways in the list.

listGatewaysResponse_httpStatus :: Lens' ListGatewaysResponse Int

The response's http status code.