Network.AWS.AlexaBusiness.ListSkills

Description

Lists all enabled skills in a specific skill group.

This operation returns paginated results.

Creating a Request

data ListSkills

See: newListSkills smart constructor.

Constructors

ListSkills'

Fields

nextToken :: Maybe Text

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

maxResults :: Maybe Natural

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

skillType :: Maybe SkillTypeFilter

Whether the skill is publicly available or is a private skill.

skillGroupArn :: Maybe Text

The ARN of the skill group for which to list enabled skills.

enablementType :: Maybe EnablementTypeFilter

Whether the skill is enabled under the user's account.

newListSkills :: ListSkills

Create a value of ListSkills with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSkills', listSkills_nextToken - An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

$sel:maxResults:ListSkills', listSkills_maxResults - The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

$sel:skillType:ListSkills', listSkills_skillType - Whether the skill is publicly available or is a private skill.

$sel:skillGroupArn:ListSkills', listSkills_skillGroupArn - The ARN of the skill group for which to list enabled skills.

$sel:enablementType:ListSkills', listSkills_enablementType - Whether the skill is enabled under the user's account.

Request Lenses

listSkills_nextToken :: Lens' ListSkills (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

listSkills_maxResults :: Lens' ListSkills (Maybe Natural)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

listSkills_skillType :: Lens' ListSkills (Maybe SkillTypeFilter)

Whether the skill is publicly available or is a private skill.

listSkills_skillGroupArn :: Lens' ListSkills (Maybe Text)

The ARN of the skill group for which to list enabled skills.

listSkills_enablementType :: Lens' ListSkills (Maybe EnablementTypeFilter)

Whether the skill is enabled under the user's account.

Destructuring the Response

data ListSkillsResponse

See: newListSkillsResponse smart constructor.

Constructors

ListSkillsResponse'

Fields

nextToken :: Maybe Text

The token returned to indicate that there is more data available.

skillSummaries :: Maybe [SkillSummary]

The list of enabled skills requested. Required.

httpStatus :: Int

The response's http status code.

newListSkillsResponse

Arguments

:: Int
$sel:httpStatus:ListSkillsResponse'

-> ListSkillsResponse

Create a value of ListSkillsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSkills', listSkillsResponse_nextToken - The token returned to indicate that there is more data available.

$sel:skillSummaries:ListSkillsResponse', listSkillsResponse_skillSummaries - The list of enabled skills requested. Required.

$sel:httpStatus:ListSkillsResponse', listSkillsResponse_httpStatus - The response's http status code.

Response Lenses

listSkillsResponse_nextToken :: Lens' ListSkillsResponse (Maybe Text)

The token returned to indicate that there is more data available.

listSkillsResponse_skillSummaries :: Lens' ListSkillsResponse (Maybe [SkillSummary])

The list of enabled skills requested. Required.

listSkillsResponse_httpStatus :: Lens' ListSkillsResponse Int

The response's http status code.