Network.AWS.AlexaBusiness.ListSkillsStoreSkillsByCategory

Description

Lists all skills in the Alexa skill store by category.

This operation returns paginated results.

Creating a Request

data ListSkillsStoreSkillsByCategory

See: newListSkillsStoreSkillsByCategory smart constructor.

Constructors

ListSkillsStoreSkillsByCategory'

Fields

nextToken :: Maybe Text

The tokens used for pagination.

maxResults :: Maybe Natural

The maximum number of skills returned per paginated calls.

categoryId :: Natural

The category ID for which the skills are being retrieved from the skill store.

newListSkillsStoreSkillsByCategory

Arguments

:: Natural
$sel:categoryId:ListSkillsStoreSkillsByCategory'

-> ListSkillsStoreSkillsByCategory

Create a value of ListSkillsStoreSkillsByCategory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSkillsStoreSkillsByCategory', listSkillsStoreSkillsByCategory_nextToken - The tokens used for pagination.

$sel:maxResults:ListSkillsStoreSkillsByCategory', listSkillsStoreSkillsByCategory_maxResults - The maximum number of skills returned per paginated calls.

$sel:categoryId:ListSkillsStoreSkillsByCategory', listSkillsStoreSkillsByCategory_categoryId - The category ID for which the skills are being retrieved from the skill store.

Request Lenses

listSkillsStoreSkillsByCategory_nextToken :: Lens' ListSkillsStoreSkillsByCategory (Maybe Text)

The tokens used for pagination.

listSkillsStoreSkillsByCategory_maxResults :: Lens' ListSkillsStoreSkillsByCategory (Maybe Natural)

The maximum number of skills returned per paginated calls.

listSkillsStoreSkillsByCategory_categoryId :: Lens' ListSkillsStoreSkillsByCategory Natural

The category ID for which the skills are being retrieved from the skill store.

Destructuring the Response

data ListSkillsStoreSkillsByCategoryResponse

See: newListSkillsStoreSkillsByCategoryResponse smart constructor.

Constructors

ListSkillsStoreSkillsByCategoryResponse'

Fields

skillsStoreSkills :: Maybe [SkillsStoreSkill]

The skill store skills.

nextToken :: Maybe Text

The tokens used for pagination.

httpStatus :: Int

The response's http status code.

newListSkillsStoreSkillsByCategoryResponse

Arguments

:: Int
$sel:httpStatus:ListSkillsStoreSkillsByCategoryResponse'

-> ListSkillsStoreSkillsByCategoryResponse

Create a value of ListSkillsStoreSkillsByCategoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:skillsStoreSkills:ListSkillsStoreSkillsByCategoryResponse', listSkillsStoreSkillsByCategoryResponse_skillsStoreSkills - The skill store skills.

$sel:nextToken:ListSkillsStoreSkillsByCategory', listSkillsStoreSkillsByCategoryResponse_nextToken - The tokens used for pagination.

$sel:httpStatus:ListSkillsStoreSkillsByCategoryResponse', listSkillsStoreSkillsByCategoryResponse_httpStatus - The response's http status code.

Response Lenses

listSkillsStoreSkillsByCategoryResponse_skillsStoreSkills :: Lens' ListSkillsStoreSkillsByCategoryResponse (Maybe [SkillsStoreSkill])

The skill store skills.

listSkillsStoreSkillsByCategoryResponse_nextToken :: Lens' ListSkillsStoreSkillsByCategoryResponse (Maybe Text)

The tokens used for pagination.

listSkillsStoreSkillsByCategoryResponse_httpStatus :: Lens' ListSkillsStoreSkillsByCategoryResponse Int

The response's http status code.