Network.AWS.AlexaBusiness.ResolveRoom

Description

Determines the details for the room from which a skill request was invoked. This operation is used by skill developers.

To query ResolveRoom from an Alexa skill, the skill ID needs to be authorized. When the skill is using an AWS Lambda function, the skill is automatically authorized when you publish your skill as a private skill to your AWS account. Skills that are hosted using a custom web service must be manually authorized. To get your skill authorized, contact AWS Support with your AWS account ID that queries the ResolveRoom API and skill ID.

Creating a Request

data ResolveRoom

See: newResolveRoom smart constructor.

Constructors

ResolveRoom'

Fields

userId :: Text

The ARN of the user. Required.

skillId :: Text

The ARN of the skill that was requested. Required.

newResolveRoom

Arguments

:: Text
$sel:userId:ResolveRoom'

-> Text
$sel:skillId:ResolveRoom'

-> ResolveRoom

Create a value of ResolveRoom with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userId:ResolveRoom', resolveRoom_userId - The ARN of the user. Required.

$sel:skillId:ResolveRoom', resolveRoom_skillId - The ARN of the skill that was requested. Required.

Request Lenses

resolveRoom_userId :: Lens' ResolveRoom Text

The ARN of the user. Required.

resolveRoom_skillId :: Lens' ResolveRoom Text

The ARN of the skill that was requested. Required.

Destructuring the Response

data ResolveRoomResponse

See: newResolveRoomResponse smart constructor.

Constructors

ResolveRoomResponse'

Fields

roomSkillParameters :: Maybe [RoomSkillParameter]

Response to get the room profile request. Required.

roomArn :: Maybe Text

The ARN of the room from which the skill request was invoked.

roomName :: Maybe Text

The name of the room from which the skill request was invoked.

httpStatus :: Int

The response's http status code.

newResolveRoomResponse

Arguments

:: Int
$sel:httpStatus:ResolveRoomResponse'

-> ResolveRoomResponse

Create a value of ResolveRoomResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roomSkillParameters:ResolveRoomResponse', resolveRoomResponse_roomSkillParameters - Response to get the room profile request. Required.

$sel:roomArn:ResolveRoomResponse', resolveRoomResponse_roomArn - The ARN of the room from which the skill request was invoked.

$sel:roomName:ResolveRoomResponse', resolveRoomResponse_roomName - The name of the room from which the skill request was invoked.

$sel:httpStatus:ResolveRoomResponse', resolveRoomResponse_httpStatus - The response's http status code.

Response Lenses

resolveRoomResponse_roomSkillParameters :: Lens' ResolveRoomResponse (Maybe [RoomSkillParameter])

Response to get the room profile request. Required.

resolveRoomResponse_roomArn :: Lens' ResolveRoomResponse (Maybe Text)

The ARN of the room from which the skill request was invoked.

resolveRoomResponse_roomName :: Lens' ResolveRoomResponse (Maybe Text)

The name of the room from which the skill request was invoked.

resolveRoomResponse_httpStatus :: Lens' ResolveRoomResponse Int

The response's http status code.