Network.AWS.AlexaBusiness.SearchAddressBooks

Description

Searches address books and lists the ones that meet a set of filter and sort criteria.

Creating a Request

data SearchAddressBooks

See: newSearchAddressBooks smart constructor.

Constructors

SearchAddressBooks'

Fields

filters :: Maybe [Filter]

The filters to use to list a specified set of address books. The supported filter key is AddressBookName.

sortCriteria :: Maybe [Sort]

The sort order to use in listing the specified set of address books. The supported sort key is AddressBookName.

nextToken :: Maybe Text

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response only includes results beyond the token, up to the value specified by MaxResults.

maxResults :: Maybe Natural

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

newSearchAddressBooks :: SearchAddressBooks

Create a value of SearchAddressBooks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:SearchAddressBooks', searchAddressBooks_filters - The filters to use to list a specified set of address books. The supported filter key is AddressBookName.

$sel:sortCriteria:SearchAddressBooks', searchAddressBooks_sortCriteria - The sort order to use in listing the specified set of address books. The supported sort key is AddressBookName.

$sel:nextToken:SearchAddressBooks', searchAddressBooks_nextToken - An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response only includes results beyond the token, up to the value specified by MaxResults.

$sel:maxResults:SearchAddressBooks', searchAddressBooks_maxResults - The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

Request Lenses

searchAddressBooks_filters :: Lens' SearchAddressBooks (Maybe [Filter])

The filters to use to list a specified set of address books. The supported filter key is AddressBookName.

searchAddressBooks_sortCriteria :: Lens' SearchAddressBooks (Maybe [Sort])

The sort order to use in listing the specified set of address books. The supported sort key is AddressBookName.

searchAddressBooks_nextToken :: Lens' SearchAddressBooks (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response only includes results beyond the token, up to the value specified by MaxResults.

searchAddressBooks_maxResults :: Lens' SearchAddressBooks (Maybe Natural)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

Destructuring the Response

data SearchAddressBooksResponse

See: newSearchAddressBooksResponse smart constructor.

Constructors

SearchAddressBooksResponse'

Fields

nextToken :: Maybe Text

The token returned to indicate that there is more data available.

addressBooks :: Maybe [AddressBookData]

The address books that meet the specified set of filter criteria, in sort order.

totalCount :: Maybe Int

The total number of address books returned.

httpStatus :: Int

The response's http status code.

newSearchAddressBooksResponse

Arguments

:: Int
$sel:httpStatus:SearchAddressBooksResponse'

-> SearchAddressBooksResponse

Create a value of SearchAddressBooksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchAddressBooks', searchAddressBooksResponse_nextToken - The token returned to indicate that there is more data available.

$sel:addressBooks:SearchAddressBooksResponse', searchAddressBooksResponse_addressBooks - The address books that meet the specified set of filter criteria, in sort order.

$sel:totalCount:SearchAddressBooksResponse', searchAddressBooksResponse_totalCount - The total number of address books returned.

$sel:httpStatus:SearchAddressBooksResponse', searchAddressBooksResponse_httpStatus - The response's http status code.

Response Lenses

searchAddressBooksResponse_nextToken :: Lens' SearchAddressBooksResponse (Maybe Text)

The token returned to indicate that there is more data available.

searchAddressBooksResponse_addressBooks :: Lens' SearchAddressBooksResponse (Maybe [AddressBookData])

The address books that meet the specified set of filter criteria, in sort order.

searchAddressBooksResponse_totalCount :: Lens' SearchAddressBooksResponse (Maybe Int)

The total number of address books returned.

searchAddressBooksResponse_httpStatus :: Lens' SearchAddressBooksResponse Int

The response's http status code.