Network.AWS.AlexaBusiness.SearchContacts

Description

Searches contacts and lists the ones that meet a set of filter and sort criteria.

Creating a Request

data SearchContacts

See: newSearchContacts smart constructor.

Constructors

SearchContacts'

Fields

nextToken :: Maybe Text

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response only includes results beyond the token, up to the value specified by MaxResults.

sortCriteria :: Maybe [Sort]

The sort order to use in listing the specified set of contacts. The supported sort keys are DisplayName, FirstName, and LastName.

maxResults :: Maybe Natural

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

filters :: Maybe [Filter]

The filters to use to list a specified set of address books. The supported filter keys are DisplayName, FirstName, LastName, and AddressBookArns.

newSearchContacts :: SearchContacts

Create a value of SearchContacts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchContacts', searchContacts_nextToken - An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response only includes results beyond the token, up to the value specified by MaxResults.

$sel:sortCriteria:SearchContacts', searchContacts_sortCriteria - The sort order to use in listing the specified set of contacts. The supported sort keys are DisplayName, FirstName, and LastName.

$sel:maxResults:SearchContacts', searchContacts_maxResults - The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

$sel:filters:SearchContacts', searchContacts_filters - The filters to use to list a specified set of address books. The supported filter keys are DisplayName, FirstName, LastName, and AddressBookArns.

Request Lenses

searchContacts_nextToken :: Lens' SearchContacts (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response only includes results beyond the token, up to the value specified by MaxResults.

searchContacts_sortCriteria :: Lens' SearchContacts (Maybe [Sort])

The sort order to use in listing the specified set of contacts. The supported sort keys are DisplayName, FirstName, and LastName.

searchContacts_maxResults :: Lens' SearchContacts (Maybe Natural)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

searchContacts_filters :: Lens' SearchContacts (Maybe [Filter])

The filters to use to list a specified set of address books. The supported filter keys are DisplayName, FirstName, LastName, and AddressBookArns.

Destructuring the Response

data SearchContactsResponse

See: newSearchContactsResponse smart constructor.

Constructors

SearchContactsResponse'

Fields

nextToken :: Maybe Text

The token returned to indicate that there is more data available.

totalCount :: Maybe Int

The total number of contacts returned.

contacts :: Maybe [ContactData]

The contacts that meet the specified set of filter criteria, in sort order.

httpStatus :: Int

The response's http status code.

newSearchContactsResponse

Arguments

:: Int
$sel:httpStatus:SearchContactsResponse'

-> SearchContactsResponse

Create a value of SearchContactsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchContacts', searchContactsResponse_nextToken - The token returned to indicate that there is more data available.

$sel:totalCount:SearchContactsResponse', searchContactsResponse_totalCount - The total number of contacts returned.

$sel:contacts:SearchContactsResponse', searchContactsResponse_contacts - The contacts that meet the specified set of filter criteria, in sort order.

$sel:httpStatus:SearchContactsResponse', searchContactsResponse_httpStatus - The response's http status code.

Response Lenses

searchContactsResponse_nextToken :: Lens' SearchContactsResponse (Maybe Text)

The token returned to indicate that there is more data available.

searchContactsResponse_totalCount :: Lens' SearchContactsResponse (Maybe Int)

The total number of contacts returned.

searchContactsResponse_contacts :: Lens' SearchContactsResponse (Maybe [ContactData])

The contacts that meet the specified set of filter criteria, in sort order.

searchContactsResponse_httpStatus :: Lens' SearchContactsResponse Int

The response's http status code.