Network.AWS.AlexaBusiness.SearchNetworkProfiles

Description

Searches network profiles and lists the ones that meet a set of filter and sort criteria.

Creating a Request

data SearchNetworkProfiles

See: newSearchNetworkProfiles smart constructor.

Constructors

SearchNetworkProfiles'

Fields

nextToken :: Maybe Text

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

sortCriteria :: Maybe [Sort]

The sort order to use to list the specified set of network profiles. Valid sort criteria includes NetworkProfileName, Ssid, and SecurityType.

maxResults :: Maybe Natural

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

filters :: Maybe [Filter]

The filters to use to list a specified set of network profiles. Valid filters are NetworkProfileName, Ssid, and SecurityType.

newSearchNetworkProfiles :: SearchNetworkProfiles

Create a value of SearchNetworkProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchNetworkProfiles', searchNetworkProfiles_nextToken - An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

$sel:sortCriteria:SearchNetworkProfiles', searchNetworkProfiles_sortCriteria - The sort order to use to list the specified set of network profiles. Valid sort criteria includes NetworkProfileName, Ssid, and SecurityType.

$sel:maxResults:SearchNetworkProfiles', searchNetworkProfiles_maxResults - The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

$sel:filters:SearchNetworkProfiles', searchNetworkProfiles_filters - The filters to use to list a specified set of network profiles. Valid filters are NetworkProfileName, Ssid, and SecurityType.

Request Lenses

searchNetworkProfiles_nextToken :: Lens' SearchNetworkProfiles (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

searchNetworkProfiles_sortCriteria :: Lens' SearchNetworkProfiles (Maybe [Sort])

The sort order to use to list the specified set of network profiles. Valid sort criteria includes NetworkProfileName, Ssid, and SecurityType.

searchNetworkProfiles_maxResults :: Lens' SearchNetworkProfiles (Maybe Natural)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

searchNetworkProfiles_filters :: Lens' SearchNetworkProfiles (Maybe [Filter])

The filters to use to list a specified set of network profiles. Valid filters are NetworkProfileName, Ssid, and SecurityType.

Destructuring the Response

data SearchNetworkProfilesResponse

See: newSearchNetworkProfilesResponse smart constructor.

Constructors

SearchNetworkProfilesResponse'

Fields

nextToken :: Maybe Text

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

networkProfiles :: Maybe [NetworkProfileData]

The network profiles that meet the specified set of filter criteria, in sort order. It is a list of NetworkProfileData objects.

totalCount :: Maybe Int

The total number of network profiles returned.

httpStatus :: Int

The response's http status code.

newSearchNetworkProfilesResponse

Arguments

:: Int
$sel:httpStatus:SearchNetworkProfilesResponse'

-> SearchNetworkProfilesResponse

Create a value of SearchNetworkProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchNetworkProfiles', searchNetworkProfilesResponse_nextToken - An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

$sel:networkProfiles:SearchNetworkProfilesResponse', searchNetworkProfilesResponse_networkProfiles - The network profiles that meet the specified set of filter criteria, in sort order. It is a list of NetworkProfileData objects.

$sel:totalCount:SearchNetworkProfilesResponse', searchNetworkProfilesResponse_totalCount - The total number of network profiles returned.

$sel:httpStatus:SearchNetworkProfilesResponse', searchNetworkProfilesResponse_httpStatus - The response's http status code.

Response Lenses

searchNetworkProfilesResponse_nextToken :: Lens' SearchNetworkProfilesResponse (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

searchNetworkProfilesResponse_networkProfiles :: Lens' SearchNetworkProfilesResponse (Maybe [NetworkProfileData])

The network profiles that meet the specified set of filter criteria, in sort order. It is a list of NetworkProfileData objects.

searchNetworkProfilesResponse_totalCount :: Lens' SearchNetworkProfilesResponse (Maybe Int)

The total number of network profiles returned.

searchNetworkProfilesResponse_httpStatus :: Lens' SearchNetworkProfilesResponse Int

The response's http status code.