Network.AWS.AlexaBusiness.SearchSkillGroups

Description

Searches skill groups and lists the ones that meet a set of filter and sort criteria.

This operation returns paginated results.

Creating a Request

data SearchSkillGroups

See: newSearchSkillGroups smart constructor.

Constructors

SearchSkillGroups'

Fields

filters :: Maybe [Filter]

The filters to use to list a specified set of skill groups. The supported filter key is SkillGroupName.

sortCriteria :: Maybe [Sort]

The sort order to use in listing the specified set of skill groups. The supported sort key is SkillGroupName.

nextToken :: Maybe Text

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults. Required.

maxResults :: Maybe Natural

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

newSearchSkillGroups :: SearchSkillGroups

Create a value of SearchSkillGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:SearchSkillGroups', searchSkillGroups_filters - The filters to use to list a specified set of skill groups. The supported filter key is SkillGroupName.

$sel:sortCriteria:SearchSkillGroups', searchSkillGroups_sortCriteria - The sort order to use in listing the specified set of skill groups. The supported sort key is SkillGroupName.

$sel:nextToken:SearchSkillGroups', searchSkillGroups_nextToken - An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults. Required.

$sel:maxResults:SearchSkillGroups', searchSkillGroups_maxResults - The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

Request Lenses

searchSkillGroups_filters :: Lens' SearchSkillGroups (Maybe [Filter])

The filters to use to list a specified set of skill groups. The supported filter key is SkillGroupName.

searchSkillGroups_sortCriteria :: Lens' SearchSkillGroups (Maybe [Sort])

The sort order to use in listing the specified set of skill groups. The supported sort key is SkillGroupName.

searchSkillGroups_nextToken :: Lens' SearchSkillGroups (Maybe Text)

An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults. Required.

searchSkillGroups_maxResults :: Lens' SearchSkillGroups (Maybe Natural)

The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

Destructuring the Response

data SearchSkillGroupsResponse

See: newSearchSkillGroupsResponse smart constructor.

Constructors

SearchSkillGroupsResponse'

Fields

nextToken :: Maybe Text

The token returned to indicate that there is more data available.

skillGroups :: Maybe [SkillGroupData]

The skill groups that meet the filter criteria, in sort order.

totalCount :: Maybe Int

The total number of skill groups returned.

httpStatus :: Int

The response's http status code.

newSearchSkillGroupsResponse

Arguments

:: Int
$sel:httpStatus:SearchSkillGroupsResponse'

-> SearchSkillGroupsResponse

Create a value of SearchSkillGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchSkillGroups', searchSkillGroupsResponse_nextToken - The token returned to indicate that there is more data available.

$sel:skillGroups:SearchSkillGroupsResponse', searchSkillGroupsResponse_skillGroups - The skill groups that meet the filter criteria, in sort order.

$sel:totalCount:SearchSkillGroupsResponse', searchSkillGroupsResponse_totalCount - The total number of skill groups returned.

$sel:httpStatus:SearchSkillGroupsResponse', searchSkillGroupsResponse_httpStatus - The response's http status code.

Response Lenses

searchSkillGroupsResponse_nextToken :: Lens' SearchSkillGroupsResponse (Maybe Text)

The token returned to indicate that there is more data available.

searchSkillGroupsResponse_skillGroups :: Lens' SearchSkillGroupsResponse (Maybe [SkillGroupData])

The skill groups that meet the filter criteria, in sort order.

searchSkillGroupsResponse_totalCount :: Lens' SearchSkillGroupsResponse (Maybe Int)

The total number of skill groups returned.

searchSkillGroupsResponse_httpStatus :: Lens' SearchSkillGroupsResponse Int

The response's http status code.