Network.AWS.AlexaBusiness.StartDeviceSync

Description

Resets a device and its account to the known default settings. This clears all information and settings set by previous users in the following ways:

Bluetooth - This unpairs all bluetooth devices paired with your echo device.

Volume - This resets the echo device's volume to the default value.

Notifications - This clears all notifications from your echo device.

Lists - This clears all to-do items from your echo device.

Settings - This internally syncs the room's profile (if the device is assigned to a room), contacts, address books, delegation access for account linking, and communications (if enabled on the room profile).

Creating a Request

data StartDeviceSync

See: newStartDeviceSync smart constructor.

Constructors

StartDeviceSync'

Fields

deviceArn :: Maybe Text

The ARN of the device to sync. Required.

roomArn :: Maybe Text

The ARN of the room with which the device to sync is associated. Required.

features :: [Feature]

Request structure to start the device sync. Required.

newStartDeviceSync :: StartDeviceSync

Create a value of StartDeviceSync with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceArn:StartDeviceSync', startDeviceSync_deviceArn - The ARN of the device to sync. Required.

$sel:roomArn:StartDeviceSync', startDeviceSync_roomArn - The ARN of the room with which the device to sync is associated. Required.

$sel:features:StartDeviceSync', startDeviceSync_features - Request structure to start the device sync. Required.

Request Lenses

startDeviceSync_deviceArn :: Lens' StartDeviceSync (Maybe Text)

The ARN of the device to sync. Required.

startDeviceSync_roomArn :: Lens' StartDeviceSync (Maybe Text)

The ARN of the room with which the device to sync is associated. Required.

startDeviceSync_features :: Lens' StartDeviceSync [Feature]

Request structure to start the device sync. Required.

Destructuring the Response

data StartDeviceSyncResponse

See: newStartDeviceSyncResponse smart constructor.

Constructors

StartDeviceSyncResponse'

Fields

httpStatus :: Int

The response's http status code.

newStartDeviceSyncResponse

Arguments

:: Int
$sel:httpStatus:StartDeviceSyncResponse'

-> StartDeviceSyncResponse

Create a value of StartDeviceSyncResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartDeviceSyncResponse', startDeviceSyncResponse_httpStatus - The response's http status code.

Response Lenses

startDeviceSyncResponse_httpStatus :: Lens' StartDeviceSyncResponse Int

The response's http status code.