Network.AWS.AlexaBusiness.Types.BusinessReportSchedule

Description

Documentation

data BusinessReportSchedule

The schedule of the usage report.

See: newBusinessReportSchedule smart constructor.

Constructors

BusinessReportSchedule'

Fields

contentRange :: Maybe BusinessReportContentRange

The content range of the reports.

format :: Maybe BusinessReportFormat

The format of the generated report (individual CSV files or zipped files of individual files).

scheduleArn :: Maybe Text

The ARN of the business report schedule.

lastBusinessReport :: Maybe BusinessReport

The details of the last business report delivery for a specified time interval.

s3KeyPrefix :: Maybe Text

The S3 key where the report is delivered.

recurrence :: Maybe BusinessReportRecurrence

The recurrence of the reports.

s3BucketName :: Maybe Text

The S3 bucket name of the output reports.

scheduleName :: Maybe Text

The name identifier of the schedule.

newBusinessReportSchedule :: BusinessReportSchedule

Create a value of BusinessReportSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentRange:BusinessReportSchedule', businessReportSchedule_contentRange - The content range of the reports.

$sel:format:BusinessReportSchedule', businessReportSchedule_format - The format of the generated report (individual CSV files or zipped files of individual files).

$sel:scheduleArn:BusinessReportSchedule', businessReportSchedule_scheduleArn - The ARN of the business report schedule.

$sel:lastBusinessReport:BusinessReportSchedule', businessReportSchedule_lastBusinessReport - The details of the last business report delivery for a specified time interval.

$sel:s3KeyPrefix:BusinessReportSchedule', businessReportSchedule_s3KeyPrefix - The S3 key where the report is delivered.

$sel:recurrence:BusinessReportSchedule', businessReportSchedule_recurrence - The recurrence of the reports.

$sel:s3BucketName:BusinessReportSchedule', businessReportSchedule_s3BucketName - The S3 bucket name of the output reports.

$sel:scheduleName:BusinessReportSchedule', businessReportSchedule_scheduleName - The name identifier of the schedule.

businessReportSchedule_contentRange :: Lens' BusinessReportSchedule (Maybe BusinessReportContentRange)

The content range of the reports.

businessReportSchedule_format :: Lens' BusinessReportSchedule (Maybe BusinessReportFormat)

The format of the generated report (individual CSV files or zipped files of individual files).

businessReportSchedule_scheduleArn :: Lens' BusinessReportSchedule (Maybe Text)

The ARN of the business report schedule.

businessReportSchedule_lastBusinessReport :: Lens' BusinessReportSchedule (Maybe BusinessReport)

The details of the last business report delivery for a specified time interval.

businessReportSchedule_s3KeyPrefix :: Lens' BusinessReportSchedule (Maybe Text)

The S3 key where the report is delivered.

businessReportSchedule_recurrence :: Lens' BusinessReportSchedule (Maybe BusinessReportRecurrence)

The recurrence of the reports.

businessReportSchedule_s3BucketName :: Lens' BusinessReportSchedule (Maybe Text)

The S3 bucket name of the output reports.

businessReportSchedule_scheduleName :: Lens' BusinessReportSchedule (Maybe Text)

The name identifier of the schedule.