Network.AWS.AlexaBusiness.Types.ContactData

Description

Documentation

data ContactData

Information related to a contact.

See: newContactData smart constructor.

Constructors

ContactData'

Fields

phoneNumber :: Maybe (Sensitive Text)

The phone number of the contact. The phone number type defaults to WORK. You can specify PhoneNumber or PhoneNumbers. We recommend that you use PhoneNumbers, which lets you specify the phone number type and multiple numbers.

phoneNumbers :: Maybe [PhoneNumber]

The list of phone numbers for the contact.

lastName :: Maybe Text

The last name of the contact, used to call the contact on the device.

firstName :: Maybe Text

The first name of the contact, used to call the contact on the device.

contactArn :: Maybe Text

The ARN of the contact.

displayName :: Maybe Text

The name of the contact to display on the console.

sipAddresses :: Maybe [SipAddress]

The list of SIP addresses for the contact.

newContactData :: ContactData

Create a value of ContactData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumber:ContactData', contactData_phoneNumber - The phone number of the contact. The phone number type defaults to WORK. You can specify PhoneNumber or PhoneNumbers. We recommend that you use PhoneNumbers, which lets you specify the phone number type and multiple numbers.

$sel:phoneNumbers:ContactData', contactData_phoneNumbers - The list of phone numbers for the contact.

$sel:lastName:ContactData', contactData_lastName - The last name of the contact, used to call the contact on the device.

$sel:firstName:ContactData', contactData_firstName - The first name of the contact, used to call the contact on the device.

$sel:contactArn:ContactData', contactData_contactArn - The ARN of the contact.

$sel:displayName:ContactData', contactData_displayName - The name of the contact to display on the console.

$sel:sipAddresses:ContactData', contactData_sipAddresses - The list of SIP addresses for the contact.

contactData_phoneNumber :: Lens' ContactData (Maybe Text)

The phone number of the contact. The phone number type defaults to WORK. You can specify PhoneNumber or PhoneNumbers. We recommend that you use PhoneNumbers, which lets you specify the phone number type and multiple numbers.

contactData_phoneNumbers :: Lens' ContactData (Maybe [PhoneNumber])

The list of phone numbers for the contact.

contactData_lastName :: Lens' ContactData (Maybe Text)

The last name of the contact, used to call the contact on the device.

contactData_firstName :: Lens' ContactData (Maybe Text)

The first name of the contact, used to call the contact on the device.

contactData_contactArn :: Lens' ContactData (Maybe Text)

The ARN of the contact.

contactData_displayName :: Lens' ContactData (Maybe Text)

The name of the contact to display on the console.

contactData_sipAddresses :: Lens' ContactData (Maybe [SipAddress])

The list of SIP addresses for the contact.