Network.AWS.AlexaBusiness.Types.CreateMeetingRoomConfiguration

Description

Documentation

data CreateMeetingRoomConfiguration

Creates meeting room settings of a room profile.

See: newCreateMeetingRoomConfiguration smart constructor.

Constructors

CreateMeetingRoomConfiguration'

Fields

roomUtilizationMetricsEnabled :: Maybe Bool

Whether room utilization metrics are enabled or not.

endOfMeetingReminder :: Maybe CreateEndOfMeetingReminder

instantBooking :: Maybe CreateInstantBooking

Settings to automatically book a room for a configured duration if it's free when joining a meeting with Alexa.

requireCheckIn :: Maybe CreateRequireCheckIn

Settings for requiring a check in when a room is reserved. Alexa can cancel a room reservation if it's not checked into to make the room available for others. Users can check in by joining the meeting with Alexa or an AVS device, or by saying “Alexa, check in.”

newCreateMeetingRoomConfiguration :: CreateMeetingRoomConfiguration

Create a value of CreateMeetingRoomConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roomUtilizationMetricsEnabled:CreateMeetingRoomConfiguration', createMeetingRoomConfiguration_roomUtilizationMetricsEnabled - Whether room utilization metrics are enabled or not.

$sel:endOfMeetingReminder:CreateMeetingRoomConfiguration', createMeetingRoomConfiguration_endOfMeetingReminder - Undocumented member.

$sel:instantBooking:CreateMeetingRoomConfiguration', createMeetingRoomConfiguration_instantBooking - Settings to automatically book a room for a configured duration if it's free when joining a meeting with Alexa.

$sel:requireCheckIn:CreateMeetingRoomConfiguration', createMeetingRoomConfiguration_requireCheckIn - Settings for requiring a check in when a room is reserved. Alexa can cancel a room reservation if it's not checked into to make the room available for others. Users can check in by joining the meeting with Alexa or an AVS device, or by saying “Alexa, check in.”

createMeetingRoomConfiguration_roomUtilizationMetricsEnabled :: Lens' CreateMeetingRoomConfiguration (Maybe Bool)

Whether room utilization metrics are enabled or not.

createMeetingRoomConfiguration_endOfMeetingReminder :: Lens' CreateMeetingRoomConfiguration (Maybe CreateEndOfMeetingReminder)

Undocumented member.

createMeetingRoomConfiguration_instantBooking :: Lens' CreateMeetingRoomConfiguration (Maybe CreateInstantBooking)

Settings to automatically book a room for a configured duration if it's free when joining a meeting with Alexa.

createMeetingRoomConfiguration_requireCheckIn :: Lens' CreateMeetingRoomConfiguration (Maybe CreateRequireCheckIn)

Settings for requiring a check in when a room is reserved. Alexa can cancel a room reservation if it's not checked into to make the room available for others. Users can check in by joining the meeting with Alexa or an AVS device, or by saying “Alexa, check in.”