Network.AWS.AlexaBusiness.Types.CreateRequireCheckIn

Description

Documentation

data CreateRequireCheckIn

Creates settings for the require check in feature that are applied to a room profile. Require check in allows a meeting room’s Alexa or AVS device to prompt the user to check in; otherwise, the room will be released.

See: newCreateRequireCheckIn smart constructor.

Constructors

CreateRequireCheckIn'

Fields

releaseAfterMinutes :: Int

Duration between 5 and 20 minutes to determine when to release the room if it's not checked into.

enabled :: Bool

Whether require check in is enabled or not.

newCreateRequireCheckIn

Arguments

:: Int
$sel:releaseAfterMinutes:CreateRequireCheckIn'

-> Bool
$sel:enabled:CreateRequireCheckIn'

-> CreateRequireCheckIn

Create a value of CreateRequireCheckIn with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:releaseAfterMinutes:CreateRequireCheckIn', createRequireCheckIn_releaseAfterMinutes - Duration between 5 and 20 minutes to determine when to release the room if it's not checked into.

$sel:enabled:CreateRequireCheckIn', createRequireCheckIn_enabled - Whether require check in is enabled or not.

createRequireCheckIn_releaseAfterMinutes :: Lens' CreateRequireCheckIn Int

Duration between 5 and 20 minutes to determine when to release the room if it's not checked into.

createRequireCheckIn_enabled :: Lens' CreateRequireCheckIn Bool

Whether require check in is enabled or not.