Network.AWS.AlexaBusiness.Types.DeveloperInfo

Description

Documentation

data DeveloperInfo

The details about the developer that published the skill.

See: newDeveloperInfo smart constructor.

Constructors

DeveloperInfo'

Fields

developerName :: Maybe Text

The name of the developer.

email :: Maybe Text

The email of the developer.

privacyPolicy :: Maybe Text

The URL of the privacy policy.

url :: Maybe Text

The website of the developer.

newDeveloperInfo :: DeveloperInfo

Create a value of DeveloperInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:developerName:DeveloperInfo', developerInfo_developerName - The name of the developer.

$sel:email:DeveloperInfo', developerInfo_email - The email of the developer.

$sel:privacyPolicy:DeveloperInfo', developerInfo_privacyPolicy - The URL of the privacy policy.

$sel:url:DeveloperInfo', developerInfo_url - The website of the developer.

developerInfo_developerName :: Lens' DeveloperInfo (Maybe Text)

The name of the developer.

developerInfo_email :: Lens' DeveloperInfo (Maybe Text)

The email of the developer.

developerInfo_privacyPolicy :: Lens' DeveloperInfo (Maybe Text)

The URL of the privacy policy.

developerInfo_url :: Lens' DeveloperInfo (Maybe Text)

The website of the developer.