Network.AWS.AlexaBusiness.Types.Device

Description

Documentation

data Device

A device with attributes.

See: newDevice smart constructor.

Constructors

Device'

Fields

deviceStatus :: Maybe DeviceStatus

The status of a device. If the status is not READY, check the DeviceStatusInfo value for details.

deviceStatusInfo :: Maybe DeviceStatusInfo

Detailed information about a device's status.

deviceArn :: Maybe Text

The ARN of a device.

macAddress :: Maybe Text

The MAC address of a device.

deviceName :: Maybe Text

The name of a device.

roomArn :: Maybe Text

The room ARN of a device.

softwareVersion :: Maybe Text

The software version of a device.

deviceType :: Maybe Text

The type of a device.

networkProfileInfo :: Maybe DeviceNetworkProfileInfo

Detailed information about a device's network profile.

deviceSerialNumber :: Maybe Text

The serial number of a device.

newDevice :: Device

Create a value of Device with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceStatus:Device', device_deviceStatus - The status of a device. If the status is not READY, check the DeviceStatusInfo value for details.

$sel:deviceStatusInfo:Device', device_deviceStatusInfo - Detailed information about a device's status.

$sel:deviceArn:Device', device_deviceArn - The ARN of a device.

$sel:macAddress:Device', device_macAddress - The MAC address of a device.

$sel:deviceName:Device', device_deviceName - The name of a device.

$sel:roomArn:Device', device_roomArn - The room ARN of a device.

$sel:softwareVersion:Device', device_softwareVersion - The software version of a device.

$sel:deviceType:Device', device_deviceType - The type of a device.

$sel:networkProfileInfo:Device', device_networkProfileInfo - Detailed information about a device's network profile.

$sel:deviceSerialNumber:Device', device_deviceSerialNumber - The serial number of a device.

device_deviceStatus :: Lens' Device (Maybe DeviceStatus)

The status of a device. If the status is not READY, check the DeviceStatusInfo value for details.

device_deviceStatusInfo :: Lens' Device (Maybe DeviceStatusInfo)

Detailed information about a device's status.

device_deviceArn :: Lens' Device (Maybe Text)

The ARN of a device.

device_macAddress :: Lens' Device (Maybe Text)

The MAC address of a device.

device_deviceName :: Lens' Device (Maybe Text)

The name of a device.

device_roomArn :: Lens' Device (Maybe Text)

The room ARN of a device.

device_softwareVersion :: Lens' Device (Maybe Text)

The software version of a device.

device_deviceType :: Lens' Device (Maybe Text)

The type of a device.

device_networkProfileInfo :: Lens' Device (Maybe DeviceNetworkProfileInfo)

Detailed information about a device's network profile.

device_deviceSerialNumber :: Lens' Device (Maybe Text)

The serial number of a device.