Network.AWS.AlexaBusiness.Types.DeviceData

Description

Documentation

data DeviceData

Device attributes.

See: newDeviceData smart constructor.

Constructors

DeviceData'

Fields

deviceStatus :: Maybe DeviceStatus

The status of a device.

networkProfileName :: Maybe Text

The name of the network profile associated with a device.

deviceStatusInfo :: Maybe DeviceStatusInfo

Detailed information about a device's status.

createdTime :: Maybe POSIX

The time (in epoch) when the device data was created.

deviceArn :: Maybe Text

The ARN of a device.

networkProfileArn :: Maybe Text

The ARN of the network profile associated with a device.

macAddress :: Maybe Text

The MAC address of a device.

deviceName :: Maybe Text

The name of a device.

roomArn :: Maybe Text

The room ARN associated with a device.

softwareVersion :: Maybe Text

The software version of a device.

deviceType :: Maybe Text

The type of a device.

roomName :: Maybe Text

The name of the room associated with a device.

deviceSerialNumber :: Maybe Text

The serial number of a device.

newDeviceData :: DeviceData

Create a value of DeviceData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceStatus:DeviceData', deviceData_deviceStatus - The status of a device.

$sel:networkProfileName:DeviceData', deviceData_networkProfileName - The name of the network profile associated with a device.

$sel:deviceStatusInfo:DeviceData', deviceData_deviceStatusInfo - Detailed information about a device's status.

$sel:createdTime:DeviceData', deviceData_createdTime - The time (in epoch) when the device data was created.

$sel:deviceArn:DeviceData', deviceData_deviceArn - The ARN of a device.

$sel:networkProfileArn:DeviceData', deviceData_networkProfileArn - The ARN of the network profile associated with a device.

$sel:macAddress:DeviceData', deviceData_macAddress - The MAC address of a device.

$sel:deviceName:DeviceData', deviceData_deviceName - The name of a device.

$sel:roomArn:DeviceData', deviceData_roomArn - The room ARN associated with a device.

$sel:softwareVersion:DeviceData', deviceData_softwareVersion - The software version of a device.

$sel:deviceType:DeviceData', deviceData_deviceType - The type of a device.

$sel:roomName:DeviceData', deviceData_roomName - The name of the room associated with a device.

$sel:deviceSerialNumber:DeviceData', deviceData_deviceSerialNumber - The serial number of a device.

deviceData_deviceStatus :: Lens' DeviceData (Maybe DeviceStatus)

The status of a device.

deviceData_networkProfileName :: Lens' DeviceData (Maybe Text)

The name of the network profile associated with a device.

deviceData_deviceStatusInfo :: Lens' DeviceData (Maybe DeviceStatusInfo)

Detailed information about a device's status.

deviceData_createdTime :: Lens' DeviceData (Maybe UTCTime)

The time (in epoch) when the device data was created.

deviceData_deviceArn :: Lens' DeviceData (Maybe Text)

The ARN of a device.

deviceData_networkProfileArn :: Lens' DeviceData (Maybe Text)

The ARN of the network profile associated with a device.

deviceData_macAddress :: Lens' DeviceData (Maybe Text)

The MAC address of a device.

deviceData_deviceName :: Lens' DeviceData (Maybe Text)

The name of a device.

deviceData_roomArn :: Lens' DeviceData (Maybe Text)

The room ARN associated with a device.

deviceData_softwareVersion :: Lens' DeviceData (Maybe Text)

The software version of a device.

deviceData_deviceType :: Lens' DeviceData (Maybe Text)

The type of a device.

deviceData_roomName :: Lens' DeviceData (Maybe Text)

The name of the room associated with a device.

deviceData_deviceSerialNumber :: Lens' DeviceData (Maybe Text)

The serial number of a device.