Network.AWS.AlexaBusiness.Types.Gateway

Description

Documentation

data Gateway

The details of the gateway.

See: newGateway smart constructor.

Constructors

Gateway'

Fields

arn :: Maybe Text

The ARN of the gateway.

name :: Maybe Text

The name of the gateway.

gatewayGroupArn :: Maybe Text

The ARN of the gateway group that the gateway is associated to.

softwareVersion :: Maybe Text

The software version of the gateway. The gateway automatically updates its software version during normal operation.

description :: Maybe Text

The description of the gateway.

newGateway :: Gateway

Create a value of Gateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Gateway', gateway_arn - The ARN of the gateway.

$sel:name:Gateway', gateway_name - The name of the gateway.

$sel:gatewayGroupArn:Gateway', gateway_gatewayGroupArn - The ARN of the gateway group that the gateway is associated to.

$sel:softwareVersion:Gateway', gateway_softwareVersion - The software version of the gateway. The gateway automatically updates its software version during normal operation.

$sel:description:Gateway', gateway_description - The description of the gateway.

gateway_arn :: Lens' Gateway (Maybe Text)

The ARN of the gateway.

gateway_name :: Lens' Gateway (Maybe Text)

The name of the gateway.

gateway_gatewayGroupArn :: Lens' Gateway (Maybe Text)

The ARN of the gateway group that the gateway is associated to.

gateway_softwareVersion :: Lens' Gateway (Maybe Text)

The software version of the gateway. The gateway automatically updates its software version during normal operation.

gateway_description :: Lens' Gateway (Maybe Text)

The description of the gateway.