Network.AWS.AlexaBusiness.Types.NetworkProfile

Description

Documentation

data NetworkProfile

The network profile associated with a device.

See: newNetworkProfile smart constructor.

Constructors

NetworkProfile'

Fields

certificateAuthorityArn :: Maybe Text

The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

trustAnchors :: Maybe (NonEmpty Text)

The root certificates of your authentication server, which is installed on your devices and used to trust your authentication server during EAP negotiation.

currentPassword :: Maybe (Sensitive Text)

The current password of the Wi-Fi network.

eapMethod :: Maybe NetworkEapMethod

The authentication standard that is used in the EAP framework. Currently, EAP_TLS is supported.

networkProfileName :: Maybe Text

The name of the network profile associated with a device.

description :: Maybe Text

Detailed information about a device's network profile.

securityType :: Maybe NetworkSecurityType

The security type of the Wi-Fi network. This can be WPA2_ENTERPRISE, WPA2_PSK, WPA_PSK, WEP, or OPEN.

nextPassword :: Maybe (Sensitive Text)

The next, or subsequent, password of the Wi-Fi network. This password is asynchronously transmitted to the device and is used when the password of the network changes to NextPassword.

networkProfileArn :: Maybe Text

The ARN of the network profile associated with a device.

ssid :: Maybe Text

The SSID of the Wi-Fi network.

newNetworkProfile :: NetworkProfile

Create a value of NetworkProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:NetworkProfile', networkProfile_certificateAuthorityArn - The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

$sel:trustAnchors:NetworkProfile', networkProfile_trustAnchors - The root certificates of your authentication server, which is installed on your devices and used to trust your authentication server during EAP negotiation.

$sel:currentPassword:NetworkProfile', networkProfile_currentPassword - The current password of the Wi-Fi network.

$sel:eapMethod:NetworkProfile', networkProfile_eapMethod - The authentication standard that is used in the EAP framework. Currently, EAP_TLS is supported.

$sel:networkProfileName:NetworkProfile', networkProfile_networkProfileName - The name of the network profile associated with a device.

$sel:description:NetworkProfile', networkProfile_description - Detailed information about a device's network profile.

$sel:securityType:NetworkProfile', networkProfile_securityType - The security type of the Wi-Fi network. This can be WPA2_ENTERPRISE, WPA2_PSK, WPA_PSK, WEP, or OPEN.

$sel:nextPassword:NetworkProfile', networkProfile_nextPassword - The next, or subsequent, password of the Wi-Fi network. This password is asynchronously transmitted to the device and is used when the password of the network changes to NextPassword.

$sel:networkProfileArn:NetworkProfile', networkProfile_networkProfileArn - The ARN of the network profile associated with a device.

$sel:ssid:NetworkProfile', networkProfile_ssid - The SSID of the Wi-Fi network.

networkProfile_certificateAuthorityArn :: Lens' NetworkProfile (Maybe Text)

The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

networkProfile_trustAnchors :: Lens' NetworkProfile (Maybe (NonEmpty Text))

The root certificates of your authentication server, which is installed on your devices and used to trust your authentication server during EAP negotiation.

networkProfile_currentPassword :: Lens' NetworkProfile (Maybe Text)

The current password of the Wi-Fi network.

networkProfile_eapMethod :: Lens' NetworkProfile (Maybe NetworkEapMethod)

The authentication standard that is used in the EAP framework. Currently, EAP_TLS is supported.

networkProfile_networkProfileName :: Lens' NetworkProfile (Maybe Text)

The name of the network profile associated with a device.

networkProfile_description :: Lens' NetworkProfile (Maybe Text)

Detailed information about a device's network profile.

networkProfile_securityType :: Lens' NetworkProfile (Maybe NetworkSecurityType)

The security type of the Wi-Fi network. This can be WPA2_ENTERPRISE, WPA2_PSK, WPA_PSK, WEP, or OPEN.

networkProfile_nextPassword :: Lens' NetworkProfile (Maybe Text)

The next, or subsequent, password of the Wi-Fi network. This password is asynchronously transmitted to the device and is used when the password of the network changes to NextPassword.

networkProfile_networkProfileArn :: Lens' NetworkProfile (Maybe Text)

The ARN of the network profile associated with a device.

networkProfile_ssid :: Lens' NetworkProfile (Maybe Text)

The SSID of the Wi-Fi network.