Network.AWS.AlexaBusiness.Types.NetworkProfileData

Description

Documentation

data NetworkProfileData

The data associated with a network profile.

See: newNetworkProfileData smart constructor.

Constructors

NetworkProfileData'

Fields

certificateAuthorityArn :: Maybe Text

The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

eapMethod :: Maybe NetworkEapMethod

The authentication standard that is used in the EAP framework. Currently, EAP_TLS is supported.

networkProfileName :: Maybe Text

The name of the network profile associated with a device.

description :: Maybe Text

Detailed information about a device's network profile.

securityType :: Maybe NetworkSecurityType

The security type of the Wi-Fi network. This can be WPA2_ENTERPRISE, WPA2_PSK, WPA_PSK, WEP, or OPEN.

networkProfileArn :: Maybe Text

The ARN of the network profile associated with a device.

ssid :: Maybe Text

The SSID of the Wi-Fi network.

newNetworkProfileData :: NetworkProfileData

Create a value of NetworkProfileData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:NetworkProfileData', networkProfileData_certificateAuthorityArn - The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

$sel:eapMethod:NetworkProfileData', networkProfileData_eapMethod - The authentication standard that is used in the EAP framework. Currently, EAP_TLS is supported.

$sel:networkProfileName:NetworkProfileData', networkProfileData_networkProfileName - The name of the network profile associated with a device.

$sel:description:NetworkProfileData', networkProfileData_description - Detailed information about a device's network profile.

$sel:securityType:NetworkProfileData', networkProfileData_securityType - The security type of the Wi-Fi network. This can be WPA2_ENTERPRISE, WPA2_PSK, WPA_PSK, WEP, or OPEN.

$sel:networkProfileArn:NetworkProfileData', networkProfileData_networkProfileArn - The ARN of the network profile associated with a device.

$sel:ssid:NetworkProfileData', networkProfileData_ssid - The SSID of the Wi-Fi network.

networkProfileData_certificateAuthorityArn :: Lens' NetworkProfileData (Maybe Text)

The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

networkProfileData_eapMethod :: Lens' NetworkProfileData (Maybe NetworkEapMethod)

The authentication standard that is used in the EAP framework. Currently, EAP_TLS is supported.

networkProfileData_networkProfileName :: Lens' NetworkProfileData (Maybe Text)

The name of the network profile associated with a device.

networkProfileData_description :: Lens' NetworkProfileData (Maybe Text)

Detailed information about a device's network profile.

networkProfileData_securityType :: Lens' NetworkProfileData (Maybe NetworkSecurityType)

The security type of the Wi-Fi network. This can be WPA2_ENTERPRISE, WPA2_PSK, WPA_PSK, WEP, or OPEN.

networkProfileData_networkProfileArn :: Lens' NetworkProfileData (Maybe Text)

The ARN of the network profile associated with a device.

networkProfileData_ssid :: Lens' NetworkProfileData (Maybe Text)

The SSID of the Wi-Fi network.