Network.AWS.AlexaBusiness.Types.Profile

Description

Documentation

data Profile

A room profile with attributes.

See: newProfile smart constructor.

Constructors

Profile'

Fields

profileName :: Maybe Text

The name of a room profile.

isDefault :: Maybe Bool

Retrieves if the profile is default or not.

address :: Maybe Text

The address of a room profile.

locale :: Maybe Text

The locale of a room profile. (This is currently available only to a limited preview audience.)

temperatureUnit :: Maybe TemperatureUnit

The temperature unit of a room profile.

addressBookArn :: Maybe Text

The ARN of the address book.

pSTNEnabled :: Maybe Bool

The PSTN setting of a room profile.

setupModeDisabled :: Maybe Bool

The setup mode of a room profile.

maxVolumeLimit :: Maybe Int

The max volume limit of a room profile.

wakeWord :: Maybe WakeWord

The wake word of a room profile.

profileArn :: Maybe Text

The ARN of a room profile.

meetingRoomConfiguration :: Maybe MeetingRoomConfiguration

Meeting room settings of a room profile.

timezone :: Maybe Text

The time zone of a room profile.

dataRetentionOptIn :: Maybe Bool

Whether data retention of the profile is enabled.

distanceUnit :: Maybe DistanceUnit

The distance unit of a room profile.

newProfile :: Profile

Create a value of Profile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profileName:Profile', profile_profileName - The name of a room profile.

$sel:isDefault:Profile', profile_isDefault - Retrieves if the profile is default or not.

$sel:address:Profile', profile_address - The address of a room profile.

$sel:locale:Profile', profile_locale - The locale of a room profile. (This is currently available only to a limited preview audience.)

$sel:temperatureUnit:Profile', profile_temperatureUnit - The temperature unit of a room profile.

$sel:addressBookArn:Profile', profile_addressBookArn - The ARN of the address book.

$sel:pSTNEnabled:Profile', profile_pSTNEnabled - The PSTN setting of a room profile.

$sel:setupModeDisabled:Profile', profile_setupModeDisabled - The setup mode of a room profile.

$sel:maxVolumeLimit:Profile', profile_maxVolumeLimit - The max volume limit of a room profile.

$sel:wakeWord:Profile', profile_wakeWord - The wake word of a room profile.

$sel:profileArn:Profile', profile_profileArn - The ARN of a room profile.

$sel:meetingRoomConfiguration:Profile', profile_meetingRoomConfiguration - Meeting room settings of a room profile.

$sel:timezone:Profile', profile_timezone - The time zone of a room profile.

$sel:dataRetentionOptIn:Profile', profile_dataRetentionOptIn - Whether data retention of the profile is enabled.

$sel:distanceUnit:Profile', profile_distanceUnit - The distance unit of a room profile.

profile_profileName :: Lens' Profile (Maybe Text)

The name of a room profile.

profile_isDefault :: Lens' Profile (Maybe Bool)

Retrieves if the profile is default or not.

profile_address :: Lens' Profile (Maybe Text)

The address of a room profile.

profile_locale :: Lens' Profile (Maybe Text)

The locale of a room profile. (This is currently available only to a limited preview audience.)

profile_temperatureUnit :: Lens' Profile (Maybe TemperatureUnit)

The temperature unit of a room profile.

profile_addressBookArn :: Lens' Profile (Maybe Text)

The ARN of the address book.

profile_pSTNEnabled :: Lens' Profile (Maybe Bool)

The PSTN setting of a room profile.

profile_setupModeDisabled :: Lens' Profile (Maybe Bool)

The setup mode of a room profile.

profile_maxVolumeLimit :: Lens' Profile (Maybe Int)

The max volume limit of a room profile.

profile_wakeWord :: Lens' Profile (Maybe WakeWord)

The wake word of a room profile.

profile_profileArn :: Lens' Profile (Maybe Text)

The ARN of a room profile.

profile_meetingRoomConfiguration :: Lens' Profile (Maybe MeetingRoomConfiguration)

Meeting room settings of a room profile.

profile_timezone :: Lens' Profile (Maybe Text)

The time zone of a room profile.

profile_dataRetentionOptIn :: Lens' Profile (Maybe Bool)

Whether data retention of the profile is enabled.

profile_distanceUnit :: Lens' Profile (Maybe DistanceUnit)

The distance unit of a room profile.