Network.AWS.AlexaBusiness.Types.RequireCheckIn

Description

Documentation

data RequireCheckIn

Settings for the require check in feature that are applied to a room profile. Require check in allows a meeting room’s Alexa or AVS device to prompt the user to check in; otherwise, the room will be released.

See: newRequireCheckIn smart constructor.

Constructors

RequireCheckIn'

Fields

releaseAfterMinutes :: Maybe Int

Duration between 5 and 20 minutes to determine when to release the room if it's not checked into.

enabled :: Maybe Bool

Whether require check in is enabled or not.

newRequireCheckIn :: RequireCheckIn

Create a value of RequireCheckIn with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:releaseAfterMinutes:RequireCheckIn', requireCheckIn_releaseAfterMinutes - Duration between 5 and 20 minutes to determine when to release the room if it's not checked into.

$sel:enabled:RequireCheckIn', requireCheckIn_enabled - Whether require check in is enabled or not.

requireCheckIn_releaseAfterMinutes :: Lens' RequireCheckIn (Maybe Int)

Duration between 5 and 20 minutes to determine when to release the room if it's not checked into.

requireCheckIn_enabled :: Lens' RequireCheckIn (Maybe Bool)

Whether require check in is enabled or not.