Network.AWS.AlexaBusiness.Types.SkillDetails

Description

Documentation

data SkillDetails

Granular information about the skill.

See: newSkillDetails smart constructor.

Constructors

SkillDetails'

Fields

newInThisVersionBulletPoints' :: Maybe [Text]

The updates added in bullet points.

skillTypes :: Maybe [Text]

The types of skills.

bulletPoints :: Maybe [Text]

The details about what the skill supports organized as bullet points.

reviews :: Maybe (HashMap Text Text)

This member has been deprecated.

The list of reviews for the skill, including Key and Value pair.

endUserLicenseAgreement :: Maybe Text

The URL of the end user license agreement.

genericKeywords :: Maybe [Text]

The generic keywords associated with the skill that can be used to find a skill.

developerInfo :: Maybe DeveloperInfo

The details about the developer that published the skill.

productDescription :: Maybe Text

The description of the product.

invocationPhrase :: Maybe Text

The phrase used to trigger the skill.

releaseDate :: Maybe Text

The date when the skill was released.

newSkillDetails :: SkillDetails

Create a value of SkillDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:newInThisVersionBulletPoints':SkillDetails', skillDetails_newInThisVersionBulletPoints - The updates added in bullet points.

$sel:skillTypes:SkillDetails', skillDetails_skillTypes - The types of skills.

$sel:bulletPoints:SkillDetails', skillDetails_bulletPoints - The details about what the skill supports organized as bullet points.

$sel:reviews:SkillDetails', skillDetails_reviews - This member has been deprecated.

The list of reviews for the skill, including Key and Value pair.

$sel:endUserLicenseAgreement:SkillDetails', skillDetails_endUserLicenseAgreement - The URL of the end user license agreement.

$sel:genericKeywords:SkillDetails', skillDetails_genericKeywords - The generic keywords associated with the skill that can be used to find a skill.

$sel:developerInfo:SkillDetails', skillDetails_developerInfo - The details about the developer that published the skill.

$sel:productDescription:SkillDetails', skillDetails_productDescription - The description of the product.

$sel:invocationPhrase:SkillDetails', skillDetails_invocationPhrase - The phrase used to trigger the skill.

$sel:releaseDate:SkillDetails', skillDetails_releaseDate - The date when the skill was released.

skillDetails_newInThisVersionBulletPoints :: Lens' SkillDetails (Maybe [Text])

The updates added in bullet points.

skillDetails_skillTypes :: Lens' SkillDetails (Maybe [Text])

The types of skills.

skillDetails_bulletPoints :: Lens' SkillDetails (Maybe [Text])

The details about what the skill supports organized as bullet points.

skillDetails_reviews :: Lens' SkillDetails (Maybe (HashMap Text Text))

This member has been deprecated.

The list of reviews for the skill, including Key and Value pair.

skillDetails_endUserLicenseAgreement :: Lens' SkillDetails (Maybe Text)

The URL of the end user license agreement.

skillDetails_genericKeywords :: Lens' SkillDetails (Maybe [Text])

The generic keywords associated with the skill that can be used to find a skill.

skillDetails_developerInfo :: Lens' SkillDetails (Maybe DeveloperInfo)

The details about the developer that published the skill.

skillDetails_productDescription :: Lens' SkillDetails (Maybe Text)

The description of the product.

skillDetails_invocationPhrase :: Lens' SkillDetails (Maybe Text)

The phrase used to trigger the skill.

skillDetails_releaseDate :: Lens' SkillDetails (Maybe Text)

The date when the skill was released.