Network.AWS.AlexaBusiness.Types.SmartHomeAppliance

Description

Documentation

data SmartHomeAppliance

A smart home appliance that can connect to a central system. Any domestic device can be a smart appliance.

See: newSmartHomeAppliance smart constructor.

Constructors

SmartHomeAppliance'

Fields

friendlyName :: Maybe Text

The friendly name of the smart home appliance.

description :: Maybe Text

The description of the smart home appliance.

manufacturerName :: Maybe Text

The name of the manufacturer of the smart home appliance.

newSmartHomeAppliance :: SmartHomeAppliance

Create a value of SmartHomeAppliance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:friendlyName:SmartHomeAppliance', smartHomeAppliance_friendlyName - The friendly name of the smart home appliance.

$sel:description:SmartHomeAppliance', smartHomeAppliance_description - The description of the smart home appliance.

$sel:manufacturerName:SmartHomeAppliance', smartHomeAppliance_manufacturerName - The name of the manufacturer of the smart home appliance.

smartHomeAppliance_friendlyName :: Lens' SmartHomeAppliance (Maybe Text)

The friendly name of the smart home appliance.

smartHomeAppliance_description :: Lens' SmartHomeAppliance (Maybe Text)

The description of the smart home appliance.

smartHomeAppliance_manufacturerName :: Lens' SmartHomeAppliance (Maybe Text)

The name of the manufacturer of the smart home appliance.