Network.AWS.AlexaBusiness.UpdateConferenceProvider

Description

Updates an existing conference provider's settings.

Creating a Request

data UpdateConferenceProvider

See: newUpdateConferenceProvider smart constructor.

Constructors

UpdateConferenceProvider'

Fields

pSTNDialIn :: Maybe PSTNDialIn

The information for PSTN conferencing.

iPDialIn :: Maybe IPDialIn

The IP endpoint and protocol for calling.

conferenceProviderArn :: Text

The ARN of the conference provider.

conferenceProviderType :: ConferenceProviderType

The type of the conference provider.

meetingSetting :: MeetingSetting

The meeting settings for the conference provider.

newUpdateConferenceProvider

Arguments

:: Text
$sel:conferenceProviderArn:UpdateConferenceProvider'

-> ConferenceProviderType
$sel:conferenceProviderType:UpdateConferenceProvider'

-> MeetingSetting
$sel:meetingSetting:UpdateConferenceProvider'

-> UpdateConferenceProvider

Create a value of UpdateConferenceProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pSTNDialIn:UpdateConferenceProvider', updateConferenceProvider_pSTNDialIn - The information for PSTN conferencing.

$sel:iPDialIn:UpdateConferenceProvider', updateConferenceProvider_iPDialIn - The IP endpoint and protocol for calling.

$sel:conferenceProviderArn:UpdateConferenceProvider', updateConferenceProvider_conferenceProviderArn - The ARN of the conference provider.

$sel:conferenceProviderType:UpdateConferenceProvider', updateConferenceProvider_conferenceProviderType - The type of the conference provider.

$sel:meetingSetting:UpdateConferenceProvider', updateConferenceProvider_meetingSetting - The meeting settings for the conference provider.

Request Lenses

updateConferenceProvider_pSTNDialIn :: Lens' UpdateConferenceProvider (Maybe PSTNDialIn)

The information for PSTN conferencing.

updateConferenceProvider_iPDialIn :: Lens' UpdateConferenceProvider (Maybe IPDialIn)

The IP endpoint and protocol for calling.

updateConferenceProvider_conferenceProviderArn :: Lens' UpdateConferenceProvider Text

The ARN of the conference provider.

updateConferenceProvider_conferenceProviderType :: Lens' UpdateConferenceProvider ConferenceProviderType

The type of the conference provider.

updateConferenceProvider_meetingSetting :: Lens' UpdateConferenceProvider MeetingSetting

The meeting settings for the conference provider.

Destructuring the Response

data UpdateConferenceProviderResponse

See: newUpdateConferenceProviderResponse smart constructor.

Constructors

UpdateConferenceProviderResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateConferenceProviderResponse

Arguments

:: Int
$sel:httpStatus:UpdateConferenceProviderResponse'

-> UpdateConferenceProviderResponse

Create a value of UpdateConferenceProviderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateConferenceProviderResponse', updateConferenceProviderResponse_httpStatus - The response's http status code.

Response Lenses

updateConferenceProviderResponse_httpStatus :: Lens' UpdateConferenceProviderResponse Int

The response's http status code.