Network.AWS.AlexaBusiness.UpdateContact

Description

Updates the contact details by the contact ARN.

Creating a Request

data UpdateContact

See: newUpdateContact smart constructor.

Constructors

UpdateContact'

Fields

phoneNumber :: Maybe (Sensitive Text)

The updated phone number of the contact. The phone number type defaults to WORK. You can either specify PhoneNumber or PhoneNumbers. We recommend that you use PhoneNumbers, which lets you specify the phone number type and multiple numbers.

phoneNumbers :: Maybe [PhoneNumber]

The list of phone numbers for the contact.

lastName :: Maybe Text

The updated last name of the contact.

firstName :: Maybe Text

The updated first name of the contact.

displayName :: Maybe Text

The updated display name of the contact.

sipAddresses :: Maybe [SipAddress]

The list of SIP addresses for the contact.

contactArn :: Text

The ARN of the contact to update.

newUpdateContact

Arguments

:: Text
$sel:contactArn:UpdateContact'

-> UpdateContact

Create a value of UpdateContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumber:UpdateContact', updateContact_phoneNumber - The updated phone number of the contact. The phone number type defaults to WORK. You can either specify PhoneNumber or PhoneNumbers. We recommend that you use PhoneNumbers, which lets you specify the phone number type and multiple numbers.

$sel:phoneNumbers:UpdateContact', updateContact_phoneNumbers - The list of phone numbers for the contact.

$sel:lastName:UpdateContact', updateContact_lastName - The updated last name of the contact.

$sel:firstName:UpdateContact', updateContact_firstName - The updated first name of the contact.

$sel:displayName:UpdateContact', updateContact_displayName - The updated display name of the contact.

$sel:sipAddresses:UpdateContact', updateContact_sipAddresses - The list of SIP addresses for the contact.

$sel:contactArn:UpdateContact', updateContact_contactArn - The ARN of the contact to update.

Request Lenses

updateContact_phoneNumber :: Lens' UpdateContact (Maybe Text)

The updated phone number of the contact. The phone number type defaults to WORK. You can either specify PhoneNumber or PhoneNumbers. We recommend that you use PhoneNumbers, which lets you specify the phone number type and multiple numbers.

updateContact_phoneNumbers :: Lens' UpdateContact (Maybe [PhoneNumber])

The list of phone numbers for the contact.

updateContact_lastName :: Lens' UpdateContact (Maybe Text)

The updated last name of the contact.

updateContact_firstName :: Lens' UpdateContact (Maybe Text)

The updated first name of the contact.

updateContact_displayName :: Lens' UpdateContact (Maybe Text)

The updated display name of the contact.

updateContact_sipAddresses :: Lens' UpdateContact (Maybe [SipAddress])

The list of SIP addresses for the contact.

updateContact_contactArn :: Lens' UpdateContact Text

The ARN of the contact to update.

Destructuring the Response

data UpdateContactResponse

See: newUpdateContactResponse smart constructor.

Constructors

UpdateContactResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateContactResponse

Arguments

:: Int
$sel:httpStatus:UpdateContactResponse'

-> UpdateContactResponse

Create a value of UpdateContactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateContactResponse', updateContactResponse_httpStatus - The response's http status code.

Response Lenses

updateContactResponse_httpStatus :: Lens' UpdateContactResponse Int

The response's http status code.