Network.AWS.AlexaBusiness.UpdateNetworkProfile

Description

Updates a network profile by the network profile ARN.

Creating a Request

data UpdateNetworkProfile

See: newUpdateNetworkProfile smart constructor.

Constructors

UpdateNetworkProfile'

Fields

certificateAuthorityArn :: Maybe Text

The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

trustAnchors :: Maybe (NonEmpty Text)

The root certificate(s) of your authentication server that will be installed on your devices and used to trust your authentication server during EAP negotiation.

currentPassword :: Maybe (Sensitive Text)

The current password of the Wi-Fi network.

networkProfileName :: Maybe Text

The name of the network profile associated with a device.

description :: Maybe Text

Detailed information about a device's network profile.

nextPassword :: Maybe (Sensitive Text)

The next, or subsequent, password of the Wi-Fi network. This password is asynchronously transmitted to the device and is used when the password of the network changes to NextPassword.

networkProfileArn :: Text

The ARN of the network profile associated with a device.

newUpdateNetworkProfile

Arguments

:: Text
$sel:networkProfileArn:UpdateNetworkProfile'

-> UpdateNetworkProfile

Create a value of UpdateNetworkProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:UpdateNetworkProfile', updateNetworkProfile_certificateAuthorityArn - The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

$sel:trustAnchors:UpdateNetworkProfile', updateNetworkProfile_trustAnchors - The root certificate(s) of your authentication server that will be installed on your devices and used to trust your authentication server during EAP negotiation.

$sel:currentPassword:UpdateNetworkProfile', updateNetworkProfile_currentPassword - The current password of the Wi-Fi network.

$sel:networkProfileName:UpdateNetworkProfile', updateNetworkProfile_networkProfileName - The name of the network profile associated with a device.

$sel:description:UpdateNetworkProfile', updateNetworkProfile_description - Detailed information about a device's network profile.

$sel:nextPassword:UpdateNetworkProfile', updateNetworkProfile_nextPassword - The next, or subsequent, password of the Wi-Fi network. This password is asynchronously transmitted to the device and is used when the password of the network changes to NextPassword.

$sel:networkProfileArn:UpdateNetworkProfile', updateNetworkProfile_networkProfileArn - The ARN of the network profile associated with a device.

Request Lenses

updateNetworkProfile_certificateAuthorityArn :: Lens' UpdateNetworkProfile (Maybe Text)

The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

updateNetworkProfile_trustAnchors :: Lens' UpdateNetworkProfile (Maybe (NonEmpty Text))

The root certificate(s) of your authentication server that will be installed on your devices and used to trust your authentication server during EAP negotiation.

updateNetworkProfile_currentPassword :: Lens' UpdateNetworkProfile (Maybe Text)

The current password of the Wi-Fi network.

updateNetworkProfile_networkProfileName :: Lens' UpdateNetworkProfile (Maybe Text)

The name of the network profile associated with a device.

updateNetworkProfile_description :: Lens' UpdateNetworkProfile (Maybe Text)

Detailed information about a device's network profile.

updateNetworkProfile_nextPassword :: Lens' UpdateNetworkProfile (Maybe Text)

The next, or subsequent, password of the Wi-Fi network. This password is asynchronously transmitted to the device and is used when the password of the network changes to NextPassword.

updateNetworkProfile_networkProfileArn :: Lens' UpdateNetworkProfile Text

The ARN of the network profile associated with a device.

Destructuring the Response

data UpdateNetworkProfileResponse

See: newUpdateNetworkProfileResponse smart constructor.

Constructors

UpdateNetworkProfileResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateNetworkProfileResponse

Arguments

:: Int
$sel:httpStatus:UpdateNetworkProfileResponse'

-> UpdateNetworkProfileResponse

Create a value of UpdateNetworkProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateNetworkProfileResponse', updateNetworkProfileResponse_httpStatus - The response's http status code.

Response Lenses

updateNetworkProfileResponse_httpStatus :: Lens' UpdateNetworkProfileResponse Int

The response's http status code.