Network.AWS.AlexaBusiness.UpdateProfile

Description

Updates an existing room profile by room profile ARN.

Creating a Request

data UpdateProfile

See: newUpdateProfile smart constructor.

Constructors

UpdateProfile'

Fields

profileName :: Maybe Text

The updated name for the room profile.

isDefault :: Maybe Bool

Sets the profile as default if selected. If this is missing, no update is done to the default status.

address :: Maybe Text

The updated address for the room profile.

locale :: Maybe Text

The updated locale for the room profile. (This is currently only available to a limited preview audience.)

temperatureUnit :: Maybe TemperatureUnit

The updated temperature unit for the room profile.

pSTNEnabled :: Maybe Bool

Whether the PSTN setting of the room profile is enabled.

setupModeDisabled :: Maybe Bool

Whether the setup mode of the profile is enabled.

maxVolumeLimit :: Maybe Int

The updated maximum volume limit for the room profile.

wakeWord :: Maybe WakeWord

The updated wake word for the room profile.

profileArn :: Maybe Text

The ARN of the room profile to update. Required.

meetingRoomConfiguration :: Maybe UpdateMeetingRoomConfiguration

The updated meeting room settings of a room profile.

timezone :: Maybe Text

The updated timezone for the room profile.

dataRetentionOptIn :: Maybe Bool

Whether data retention of the profile is enabled.

distanceUnit :: Maybe DistanceUnit

The updated distance unit for the room profile.

newUpdateProfile :: UpdateProfile

Create a value of UpdateProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:profileName:UpdateProfile', updateProfile_profileName - The updated name for the room profile.

$sel:isDefault:UpdateProfile', updateProfile_isDefault - Sets the profile as default if selected. If this is missing, no update is done to the default status.

$sel:address:UpdateProfile', updateProfile_address - The updated address for the room profile.

$sel:locale:UpdateProfile', updateProfile_locale - The updated locale for the room profile. (This is currently only available to a limited preview audience.)

$sel:temperatureUnit:UpdateProfile', updateProfile_temperatureUnit - The updated temperature unit for the room profile.

$sel:pSTNEnabled:UpdateProfile', updateProfile_pSTNEnabled - Whether the PSTN setting of the room profile is enabled.

$sel:setupModeDisabled:UpdateProfile', updateProfile_setupModeDisabled - Whether the setup mode of the profile is enabled.

$sel:maxVolumeLimit:UpdateProfile', updateProfile_maxVolumeLimit - The updated maximum volume limit for the room profile.

$sel:wakeWord:UpdateProfile', updateProfile_wakeWord - The updated wake word for the room profile.

$sel:profileArn:UpdateProfile', updateProfile_profileArn - The ARN of the room profile to update. Required.

$sel:meetingRoomConfiguration:UpdateProfile', updateProfile_meetingRoomConfiguration - The updated meeting room settings of a room profile.

$sel:timezone:UpdateProfile', updateProfile_timezone - The updated timezone for the room profile.

$sel:dataRetentionOptIn:UpdateProfile', updateProfile_dataRetentionOptIn - Whether data retention of the profile is enabled.

$sel:distanceUnit:UpdateProfile', updateProfile_distanceUnit - The updated distance unit for the room profile.

Request Lenses

updateProfile_profileName :: Lens' UpdateProfile (Maybe Text)

The updated name for the room profile.

updateProfile_isDefault :: Lens' UpdateProfile (Maybe Bool)

Sets the profile as default if selected. If this is missing, no update is done to the default status.

updateProfile_address :: Lens' UpdateProfile (Maybe Text)

The updated address for the room profile.

updateProfile_locale :: Lens' UpdateProfile (Maybe Text)

The updated locale for the room profile. (This is currently only available to a limited preview audience.)

updateProfile_temperatureUnit :: Lens' UpdateProfile (Maybe TemperatureUnit)

The updated temperature unit for the room profile.

updateProfile_pSTNEnabled :: Lens' UpdateProfile (Maybe Bool)

Whether the PSTN setting of the room profile is enabled.

updateProfile_setupModeDisabled :: Lens' UpdateProfile (Maybe Bool)

Whether the setup mode of the profile is enabled.

updateProfile_maxVolumeLimit :: Lens' UpdateProfile (Maybe Int)

The updated maximum volume limit for the room profile.

updateProfile_wakeWord :: Lens' UpdateProfile (Maybe WakeWord)

The updated wake word for the room profile.

updateProfile_profileArn :: Lens' UpdateProfile (Maybe Text)

The ARN of the room profile to update. Required.

updateProfile_meetingRoomConfiguration :: Lens' UpdateProfile (Maybe UpdateMeetingRoomConfiguration)

The updated meeting room settings of a room profile.

updateProfile_timezone :: Lens' UpdateProfile (Maybe Text)

The updated timezone for the room profile.

updateProfile_dataRetentionOptIn :: Lens' UpdateProfile (Maybe Bool)

Whether data retention of the profile is enabled.

updateProfile_distanceUnit :: Lens' UpdateProfile (Maybe DistanceUnit)

The updated distance unit for the room profile.

Destructuring the Response

data UpdateProfileResponse

See: newUpdateProfileResponse smart constructor.

Constructors

UpdateProfileResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateProfileResponse

Arguments

:: Int
$sel:httpStatus:UpdateProfileResponse'

-> UpdateProfileResponse

Create a value of UpdateProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateProfileResponse', updateProfileResponse_httpStatus - The response's http status code.

Response Lenses

updateProfileResponse_httpStatus :: Lens' UpdateProfileResponse Int

The response's http status code.