Network.AWS.AlexaBusiness.UpdateRoom

Description

Updates room details by room ARN.

Creating a Request

data UpdateRoom

See: newUpdateRoom smart constructor.

Constructors

UpdateRoom'

Fields

roomArn :: Maybe Text

The ARN of the room to update.

profileArn :: Maybe Text

The updated profile ARN for the room.

description :: Maybe Text

The updated description for the room.

providerCalendarId :: Maybe Text

The updated provider calendar ARN for the room.

roomName :: Maybe Text

The updated name for the room.

newUpdateRoom :: UpdateRoom

Create a value of UpdateRoom with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roomArn:UpdateRoom', updateRoom_roomArn - The ARN of the room to update.

$sel:profileArn:UpdateRoom', updateRoom_profileArn - The updated profile ARN for the room.

$sel:description:UpdateRoom', updateRoom_description - The updated description for the room.

$sel:providerCalendarId:UpdateRoom', updateRoom_providerCalendarId - The updated provider calendar ARN for the room.

$sel:roomName:UpdateRoom', updateRoom_roomName - The updated name for the room.

Request Lenses

updateRoom_roomArn :: Lens' UpdateRoom (Maybe Text)

The ARN of the room to update.

updateRoom_profileArn :: Lens' UpdateRoom (Maybe Text)

The updated profile ARN for the room.

updateRoom_description :: Lens' UpdateRoom (Maybe Text)

The updated description for the room.

updateRoom_providerCalendarId :: Lens' UpdateRoom (Maybe Text)

The updated provider calendar ARN for the room.

updateRoom_roomName :: Lens' UpdateRoom (Maybe Text)

The updated name for the room.

Destructuring the Response

data UpdateRoomResponse

See: newUpdateRoomResponse smart constructor.

Constructors

UpdateRoomResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateRoomResponse

Arguments

:: Int
$sel:httpStatus:UpdateRoomResponse'

-> UpdateRoomResponse

Create a value of UpdateRoomResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateRoomResponse', updateRoomResponse_httpStatus - The response's http status code.

Response Lenses

updateRoomResponse_httpStatus :: Lens' UpdateRoomResponse Int

The response's http status code.