Network.AWS.Amplify.CreateApp

Description

Creates a new Amplify app.

Creating a Request

data CreateApp

The request structure used to create apps in Amplify.

See: newCreateApp smart constructor.

Constructors

CreateApp'

Fields

enableBranchAutoBuild :: Maybe Bool

Enables the auto building of branches for an Amplify app.

oauthToken :: Maybe (Sensitive Text)

The OAuth token for a third-party source control system for an Amplify app. The OAuth token is used to create a webhook and a read-only deploy key. The OAuth token is not stored.

accessToken :: Maybe (Sensitive Text)

The personal access token for a third-party source control system for an Amplify app. The personal access token is used to create a webhook and a read-only deploy key. The token is not stored.

customHeaders :: Maybe Text

The custom HTTP headers for an Amplify app.

platform :: Maybe Platform

The platform or framework for an Amplify app.

basicAuthCredentials :: Maybe (Sensitive Text)

The credentials for basic authorization for an Amplify app.

repository :: Maybe Text

The repository for an Amplify app.

buildSpec :: Maybe Text

The build specification (build spec) for an Amplify app.

enableBranchAutoDeletion :: Maybe Bool

Automatically disconnects a branch in the Amplify Console when you delete a branch from your Git repository.

customRules :: Maybe [CustomRule]

The custom rewrite and redirect rules for an Amplify app.

iamServiceRoleArn :: Maybe Text

The AWS Identity and Access Management (IAM) service role for an Amplify app.

autoBranchCreationPatterns :: Maybe [Text]

The automated branch creation glob patterns for an Amplify app.

autoBranchCreationConfig :: Maybe AutoBranchCreationConfig

The automated branch creation configuration for an Amplify app.

environmentVariables :: Maybe (HashMap Text Text)

The environment variables map for an Amplify app.

enableAutoBranchCreation :: Maybe Bool

Enables automated branch creation for an Amplify app.

enableBasicAuth :: Maybe Bool

Enables basic authorization for an Amplify app. This will apply to all branches that are part of this app.

description :: Maybe Text

The description for an Amplify app.

tags :: Maybe (HashMap Text Text)

The tag for an Amplify app.

name :: Text

The name for an Amplify app.

newCreateApp

Arguments

:: Text
$sel:name:CreateApp'

-> CreateApp

Create a value of CreateApp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enableBranchAutoBuild:CreateApp', createApp_enableBranchAutoBuild - Enables the auto building of branches for an Amplify app.

$sel:oauthToken:CreateApp', createApp_oauthToken - The OAuth token for a third-party source control system for an Amplify app. The OAuth token is used to create a webhook and a read-only deploy key. The OAuth token is not stored.

$sel:accessToken:CreateApp', createApp_accessToken - The personal access token for a third-party source control system for an Amplify app. The personal access token is used to create a webhook and a read-only deploy key. The token is not stored.

$sel:customHeaders:CreateApp', createApp_customHeaders - The custom HTTP headers for an Amplify app.

$sel:platform:CreateApp', createApp_platform - The platform or framework for an Amplify app.

$sel:basicAuthCredentials:CreateApp', createApp_basicAuthCredentials - The credentials for basic authorization for an Amplify app.

$sel:repository:CreateApp', createApp_repository - The repository for an Amplify app.

$sel:buildSpec:CreateApp', createApp_buildSpec - The build specification (build spec) for an Amplify app.

$sel:enableBranchAutoDeletion:CreateApp', createApp_enableBranchAutoDeletion - Automatically disconnects a branch in the Amplify Console when you delete a branch from your Git repository.

$sel:customRules:CreateApp', createApp_customRules - The custom rewrite and redirect rules for an Amplify app.

$sel:iamServiceRoleArn:CreateApp', createApp_iamServiceRoleArn - The AWS Identity and Access Management (IAM) service role for an Amplify app.

$sel:autoBranchCreationPatterns:CreateApp', createApp_autoBranchCreationPatterns - The automated branch creation glob patterns for an Amplify app.

$sel:autoBranchCreationConfig:CreateApp', createApp_autoBranchCreationConfig - The automated branch creation configuration for an Amplify app.

$sel:environmentVariables:CreateApp', createApp_environmentVariables - The environment variables map for an Amplify app.

$sel:enableAutoBranchCreation:CreateApp', createApp_enableAutoBranchCreation - Enables automated branch creation for an Amplify app.

$sel:enableBasicAuth:CreateApp', createApp_enableBasicAuth - Enables basic authorization for an Amplify app. This will apply to all branches that are part of this app.

$sel:description:CreateApp', createApp_description - The description for an Amplify app.

$sel:tags:CreateApp', createApp_tags - The tag for an Amplify app.

$sel:name:CreateApp', createApp_name - The name for an Amplify app.

Request Lenses

createApp_enableBranchAutoBuild :: Lens' CreateApp (Maybe Bool)

Enables the auto building of branches for an Amplify app.

createApp_oauthToken :: Lens' CreateApp (Maybe Text)

The OAuth token for a third-party source control system for an Amplify app. The OAuth token is used to create a webhook and a read-only deploy key. The OAuth token is not stored.

createApp_accessToken :: Lens' CreateApp (Maybe Text)

The personal access token for a third-party source control system for an Amplify app. The personal access token is used to create a webhook and a read-only deploy key. The token is not stored.

createApp_customHeaders :: Lens' CreateApp (Maybe Text)

The custom HTTP headers for an Amplify app.

createApp_platform :: Lens' CreateApp (Maybe Platform)

The platform or framework for an Amplify app.

createApp_basicAuthCredentials :: Lens' CreateApp (Maybe Text)

The credentials for basic authorization for an Amplify app.

createApp_repository :: Lens' CreateApp (Maybe Text)

The repository for an Amplify app.

createApp_buildSpec :: Lens' CreateApp (Maybe Text)

The build specification (build spec) for an Amplify app.

createApp_enableBranchAutoDeletion :: Lens' CreateApp (Maybe Bool)

Automatically disconnects a branch in the Amplify Console when you delete a branch from your Git repository.

createApp_customRules :: Lens' CreateApp (Maybe [CustomRule])

The custom rewrite and redirect rules for an Amplify app.

createApp_iamServiceRoleArn :: Lens' CreateApp (Maybe Text)

The AWS Identity and Access Management (IAM) service role for an Amplify app.

createApp_autoBranchCreationPatterns :: Lens' CreateApp (Maybe [Text])

The automated branch creation glob patterns for an Amplify app.

createApp_autoBranchCreationConfig :: Lens' CreateApp (Maybe AutoBranchCreationConfig)

The automated branch creation configuration for an Amplify app.

createApp_environmentVariables :: Lens' CreateApp (Maybe (HashMap Text Text))

The environment variables map for an Amplify app.

createApp_enableAutoBranchCreation :: Lens' CreateApp (Maybe Bool)

Enables automated branch creation for an Amplify app.

createApp_enableBasicAuth :: Lens' CreateApp (Maybe Bool)

Enables basic authorization for an Amplify app. This will apply to all branches that are part of this app.

createApp_description :: Lens' CreateApp (Maybe Text)

The description for an Amplify app.

createApp_tags :: Lens' CreateApp (Maybe (HashMap Text Text))

The tag for an Amplify app.

createApp_name :: Lens' CreateApp Text

The name for an Amplify app.

Destructuring the Response

data CreateAppResponse

See: newCreateAppResponse smart constructor.

Constructors

CreateAppResponse'

Fields

httpStatus :: Int

The response's http status code.

app :: App

newCreateAppResponse

Arguments

:: Int
$sel:httpStatus:CreateAppResponse'

-> App
$sel:app:CreateAppResponse'

-> CreateAppResponse

Create a value of CreateAppResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateAppResponse', createAppResponse_httpStatus - The response's http status code.

$sel:app:CreateAppResponse', createAppResponse_app - Undocumented member.

Response Lenses

createAppResponse_httpStatus :: Lens' CreateAppResponse Int

The response's http status code.

createAppResponse_app :: Lens' CreateAppResponse App

Undocumented member.