Network.AWS.Amplify.GetBranch

Description

Returns a branch for an Amplify app.

Creating a Request

data GetBranch

The request structure for the get branch request.

See: newGetBranch smart constructor.

Constructors

GetBranch'

Fields

appId :: Text

The unique ID for an Amplify app.

branchName :: Text

The name for the branch.

newGetBranch

Arguments

:: Text
$sel:appId:GetBranch'

-> Text
$sel:branchName:GetBranch'

-> GetBranch

Create a value of GetBranch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:GetBranch', getBranch_appId - The unique ID for an Amplify app.

$sel:branchName:GetBranch', getBranch_branchName - The name for the branch.

Request Lenses

getBranch_appId :: Lens' GetBranch Text

The unique ID for an Amplify app.

getBranch_branchName :: Lens' GetBranch Text

The name for the branch.

Destructuring the Response

data GetBranchResponse

See: newGetBranchResponse smart constructor.

Constructors

GetBranchResponse'

Fields

httpStatus :: Int

The response's http status code.

branch :: Branch

newGetBranchResponse

Arguments

:: Int
$sel:httpStatus:GetBranchResponse'

-> Branch
$sel:branch:GetBranchResponse'

-> GetBranchResponse

Create a value of GetBranchResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetBranchResponse', getBranchResponse_httpStatus - The response's http status code.

$sel:branch:GetBranchResponse', getBranchResponse_branch - Undocumented member.

Response Lenses

getBranchResponse_httpStatus :: Lens' GetBranchResponse Int

The response's http status code.

getBranchResponse_branch :: Lens' GetBranchResponse Branch

Undocumented member.