Network.AWS.Amplify.ListDomainAssociations

Description

Returns the domain associations for an Amplify app.

This operation returns paginated results.

Creating a Request

data ListDomainAssociations

The request structure for the list domain associations request.

See: newListDomainAssociations smart constructor.

Constructors

ListDomainAssociations'

Fields

nextToken :: Maybe Text

A pagination token. Set to null to start listing apps from the start. If non-null, a pagination token is returned in a result. Pass its value in here to list more projects.

maxResults :: Maybe Natural

The maximum number of records to list in a single response.

appId :: Text

The unique ID for an Amplify app.

newListDomainAssociations

Arguments

:: Text
$sel:appId:ListDomainAssociations'

-> ListDomainAssociations

Create a value of ListDomainAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDomainAssociations', listDomainAssociations_nextToken - A pagination token. Set to null to start listing apps from the start. If non-null, a pagination token is returned in a result. Pass its value in here to list more projects.

$sel:maxResults:ListDomainAssociations', listDomainAssociations_maxResults - The maximum number of records to list in a single response.

$sel:appId:ListDomainAssociations', listDomainAssociations_appId - The unique ID for an Amplify app.

Request Lenses

listDomainAssociations_nextToken :: Lens' ListDomainAssociations (Maybe Text)

A pagination token. Set to null to start listing apps from the start. If non-null, a pagination token is returned in a result. Pass its value in here to list more projects.

listDomainAssociations_maxResults :: Lens' ListDomainAssociations (Maybe Natural)

The maximum number of records to list in a single response.

listDomainAssociations_appId :: Lens' ListDomainAssociations Text

The unique ID for an Amplify app.

Destructuring the Response

data ListDomainAssociationsResponse

The result structure for the list domain association request.

See: newListDomainAssociationsResponse smart constructor.

Constructors

ListDomainAssociationsResponse'

Fields

nextToken :: Maybe Text

A pagination token. If non-null, a pagination token is returned in a result. Pass its value in another request to retrieve more entries.

httpStatus :: Int

The response's http status code.

domainAssociations :: [DomainAssociation]

A list of domain associations.

newListDomainAssociationsResponse

Arguments

:: Int
$sel:httpStatus:ListDomainAssociationsResponse'

-> ListDomainAssociationsResponse

Create a value of ListDomainAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDomainAssociations', listDomainAssociationsResponse_nextToken - A pagination token. If non-null, a pagination token is returned in a result. Pass its value in another request to retrieve more entries.

$sel:httpStatus:ListDomainAssociationsResponse', listDomainAssociationsResponse_httpStatus - The response's http status code.

$sel:domainAssociations:ListDomainAssociationsResponse', listDomainAssociationsResponse_domainAssociations - A list of domain associations.

Response Lenses

listDomainAssociationsResponse_nextToken :: Lens' ListDomainAssociationsResponse (Maybe Text)

A pagination token. If non-null, a pagination token is returned in a result. Pass its value in another request to retrieve more entries.

listDomainAssociationsResponse_httpStatus :: Lens' ListDomainAssociationsResponse Int

The response's http status code.

listDomainAssociationsResponse_domainAssociations :: Lens' ListDomainAssociationsResponse [DomainAssociation]

A list of domain associations.