Network.AWS.Amplify.Types

Description

Service Configuration

defaultService :: Service

API version 2017-07-25 of the Amazon Amplify SDK configuration.

Errors

_DependentServiceFailureException :: AsError a => Getting (First ServiceError) a ServiceError

An operation failed because a dependent service threw an exception.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

An entity was not found during an operation.

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The service failed to perform an operation due to an internal issue.

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

An operation failed due to a lack of access.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

An operation failed due to a non-existent resource.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

A request contains unexpected data.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A resource could not be created because service quotas were exceeded.

DomainStatus

newtype DomainStatus

Constructors

DomainStatus'

Fields

fromDomainStatus :: Text

Bundled Patterns

pattern DomainStatus_AVAILABLE :: DomainStatus

pattern DomainStatus_CREATING :: DomainStatus

pattern DomainStatus_FAILED :: DomainStatus

pattern DomainStatus_IN_PROGRESS :: DomainStatus

pattern DomainStatus_PENDING_DEPLOYMENT :: DomainStatus

pattern DomainStatus_PENDING_VERIFICATION :: DomainStatus

pattern DomainStatus_REQUESTING_CERTIFICATE :: DomainStatus

pattern DomainStatus_UPDATING :: DomainStatus

JobStatus

newtype JobStatus

Constructors

JobStatus'

Fields

fromJobStatus :: Text

Bundled Patterns

pattern JobStatus_CANCELLED :: JobStatus

pattern JobStatus_CANCELLING :: JobStatus

pattern JobStatus_FAILED :: JobStatus

pattern JobStatus_PENDING :: JobStatus

pattern JobStatus_PROVISIONING :: JobStatus

pattern JobStatus_RUNNING :: JobStatus

pattern JobStatus_SUCCEED :: JobStatus

JobType

newtype JobType

Constructors

JobType'

Fields

fromJobType :: Text

Bundled Patterns

pattern JobType_MANUAL :: JobType

pattern JobType_RELEASE :: JobType

pattern JobType_RETRY :: JobType

pattern JobType_WEB_HOOK :: JobType

Platform

newtype Platform

Constructors

Platform'

Fields

fromPlatform :: Text

Bundled Patterns

pattern Platform_WEB :: Platform

Stage

newtype Stage

Constructors

Stage'

Fields

fromStage :: Text

Bundled Patterns

pattern Stage_BETA :: Stage

pattern Stage_DEVELOPMENT :: Stage

pattern Stage_EXPERIMENTAL :: Stage

pattern Stage_PRODUCTION :: Stage

pattern Stage_PULL_REQUEST :: Stage

App

data App

Represents the different branches of a repository for building, deploying, and hosting an Amplify app.

See: newApp smart constructor.

Constructors

App'

Fields

customHeaders :: Maybe Text

Describes the custom HTTP headers for the Amplify app.

basicAuthCredentials :: Maybe (Sensitive Text)

The basic authorization credentials for branches for the Amplify app.

buildSpec :: Maybe Text

Describes the content of the build specification (build spec) for the Amplify app.

enableBranchAutoDeletion :: Maybe Bool

Automatically disconnect a branch in the Amplify Console when you delete a branch from your Git repository.

customRules :: Maybe [CustomRule]

Describes the custom redirect and rewrite rules for the Amplify app.

iamServiceRoleArn :: Maybe Text

The AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) of the Amplify app.

autoBranchCreationPatterns :: Maybe [Text]

Describes the automated branch creation glob patterns for the Amplify app.

productionBranch :: Maybe ProductionBranch

Describes the information about a production branch of the Amplify app.

autoBranchCreationConfig :: Maybe AutoBranchCreationConfig

Describes the automated branch creation configuration for the Amplify app.

enableAutoBranchCreation :: Maybe Bool

Enables automated branch creation for the Amplify app.

tags :: Maybe (HashMap Text Text)

The tag for the Amplify app.

appId :: Text

The unique ID of the Amplify app.

appArn :: Text

The Amazon Resource Name (ARN) of the Amplify app.

name :: Text

The name for the Amplify app.

description :: Text

The description for the Amplify app.

repository :: Text

The repository for the Amplify app.

platform :: Platform

The platform for the Amplify app.

createTime :: POSIX

Creates a date and time for the Amplify app.

updateTime :: POSIX

Updates the date and time for the Amplify app.

environmentVariables :: HashMap Text Text

The environment variables for the Amplify app.

defaultDomain :: Text

The default domain for the Amplify app.

enableBranchAutoBuild :: Bool

Enables the auto-building of branches for the Amplify app.

enableBasicAuth :: Bool

Enables basic authorization for the Amplify app's branches.

newApp

Arguments

:: Text
$sel:appId:App'

-> Text
$sel:appArn:App'

-> Text
$sel:name:App'

-> Text
$sel:description:App'

-> Text
$sel:repository:App'

-> Platform
$sel:platform:App'

-> UTCTime
$sel:createTime:App'

-> UTCTime
$sel:updateTime:App'

-> Text
$sel:defaultDomain:App'

-> Bool
$sel:enableBranchAutoBuild:App'

-> Bool
$sel:enableBasicAuth:App'

-> App

Create a value of App with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customHeaders:App', app_customHeaders - Describes the custom HTTP headers for the Amplify app.

$sel:basicAuthCredentials:App', app_basicAuthCredentials - The basic authorization credentials for branches for the Amplify app.

$sel:buildSpec:App', app_buildSpec - Describes the content of the build specification (build spec) for the Amplify app.

$sel:enableBranchAutoDeletion:App', app_enableBranchAutoDeletion - Automatically disconnect a branch in the Amplify Console when you delete a branch from your Git repository.

$sel:customRules:App', app_customRules - Describes the custom redirect and rewrite rules for the Amplify app.

$sel:iamServiceRoleArn:App', app_iamServiceRoleArn - The AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) of the Amplify app.

$sel:autoBranchCreationPatterns:App', app_autoBranchCreationPatterns - Describes the automated branch creation glob patterns for the Amplify app.

$sel:productionBranch:App', app_productionBranch - Describes the information about a production branch of the Amplify app.

$sel:autoBranchCreationConfig:App', app_autoBranchCreationConfig - Describes the automated branch creation configuration for the Amplify app.

$sel:enableAutoBranchCreation:App', app_enableAutoBranchCreation - Enables automated branch creation for the Amplify app.

$sel:tags:App', app_tags - The tag for the Amplify app.

$sel:appId:App', app_appId - The unique ID of the Amplify app.

$sel:appArn:App', app_appArn - The Amazon Resource Name (ARN) of the Amplify app.

$sel:name:App', app_name - The name for the Amplify app.

$sel:description:App', app_description - The description for the Amplify app.

$sel:repository:App', app_repository - The repository for the Amplify app.

$sel:platform:App', app_platform - The platform for the Amplify app.

$sel:createTime:App', app_createTime - Creates a date and time for the Amplify app.

$sel:updateTime:App', app_updateTime - Updates the date and time for the Amplify app.

$sel:environmentVariables:App', app_environmentVariables - The environment variables for the Amplify app.

$sel:defaultDomain:App', app_defaultDomain - The default domain for the Amplify app.

$sel:enableBranchAutoBuild:App', app_enableBranchAutoBuild - Enables the auto-building of branches for the Amplify app.

$sel:enableBasicAuth:App', app_enableBasicAuth - Enables basic authorization for the Amplify app's branches.

app_customHeaders :: Lens' App (Maybe Text)

Describes the custom HTTP headers for the Amplify app.

app_basicAuthCredentials :: Lens' App (Maybe Text)

The basic authorization credentials for branches for the Amplify app.

app_buildSpec :: Lens' App (Maybe Text)

Describes the content of the build specification (build spec) for the Amplify app.

app_enableBranchAutoDeletion :: Lens' App (Maybe Bool)

Automatically disconnect a branch in the Amplify Console when you delete a branch from your Git repository.

app_customRules :: Lens' App (Maybe [CustomRule])

Describes the custom redirect and rewrite rules for the Amplify app.

app_iamServiceRoleArn :: Lens' App (Maybe Text)

The AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) of the Amplify app.

app_autoBranchCreationPatterns :: Lens' App (Maybe [Text])

Describes the automated branch creation glob patterns for the Amplify app.

app_productionBranch :: Lens' App (Maybe ProductionBranch)

Describes the information about a production branch of the Amplify app.

app_autoBranchCreationConfig :: Lens' App (Maybe AutoBranchCreationConfig)

Describes the automated branch creation configuration for the Amplify app.

app_enableAutoBranchCreation :: Lens' App (Maybe Bool)

Enables automated branch creation for the Amplify app.

app_tags :: Lens' App (Maybe (HashMap Text Text))

The tag for the Amplify app.

app_appId :: Lens' App Text

The unique ID of the Amplify app.

app_appArn :: Lens' App Text

The Amazon Resource Name (ARN) of the Amplify app.

app_name :: Lens' App Text

The name for the Amplify app.

app_description :: Lens' App Text

The description for the Amplify app.

app_repository :: Lens' App Text

The repository for the Amplify app.

app_platform :: Lens' App Platform

The platform for the Amplify app.

app_createTime :: Lens' App UTCTime

Creates a date and time for the Amplify app.

app_updateTime :: Lens' App UTCTime

Updates the date and time for the Amplify app.

app_environmentVariables :: Lens' App (HashMap Text Text)

The environment variables for the Amplify app.

app_defaultDomain :: Lens' App Text

The default domain for the Amplify app.

app_enableBranchAutoBuild :: Lens' App Bool

Enables the auto-building of branches for the Amplify app.

app_enableBasicAuth :: Lens' App Bool

Enables basic authorization for the Amplify app's branches.

Artifact

data Artifact

Describes an artifact.

See: newArtifact smart constructor.

Constructors

Artifact'

Fields

artifactFileName :: Text

The file name for the artifact.

artifactId :: Text

The unique ID for the artifact.

newArtifact

Arguments

:: Text
$sel:artifactFileName:Artifact'

-> Text
$sel:artifactId:Artifact'

-> Artifact

Create a value of Artifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:artifactFileName:Artifact', artifact_artifactFileName - The file name for the artifact.

$sel:artifactId:Artifact', artifact_artifactId - The unique ID for the artifact.

artifact_artifactFileName :: Lens' Artifact Text

The file name for the artifact.

artifact_artifactId :: Lens' Artifact Text

The unique ID for the artifact.

AutoBranchCreationConfig

data AutoBranchCreationConfig

Describes the automated branch creation configuration.

See: newAutoBranchCreationConfig smart constructor.

Constructors

AutoBranchCreationConfig'

Fields

framework :: Maybe Text

The framework for the autocreated branch.

stage :: Maybe Stage

Describes the current stage for the autocreated branch.

enablePullRequestPreview :: Maybe Bool

Enables pull request previews for the autocreated branch.

basicAuthCredentials :: Maybe (Sensitive Text)

The basic authorization credentials for the autocreated branch.

buildSpec :: Maybe Text

The build specification (build spec) for the autocreated branch.

enablePerformanceMode :: Maybe Bool

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

environmentVariables :: Maybe (HashMap Text Text)

The environment variables for the autocreated branch.

enableAutoBuild :: Maybe Bool

Enables auto building for the autocreated branch.

enableBasicAuth :: Maybe Bool

Enables basic authorization for the autocreated branch.

pullRequestEnvironmentName :: Maybe Text

The Amplify environment name for the pull request.

newAutoBranchCreationConfig :: AutoBranchCreationConfig

Create a value of AutoBranchCreationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:framework:AutoBranchCreationConfig', autoBranchCreationConfig_framework - The framework for the autocreated branch.

$sel:stage:AutoBranchCreationConfig', autoBranchCreationConfig_stage - Describes the current stage for the autocreated branch.

$sel:enablePullRequestPreview:AutoBranchCreationConfig', autoBranchCreationConfig_enablePullRequestPreview - Enables pull request previews for the autocreated branch.

$sel:basicAuthCredentials:AutoBranchCreationConfig', autoBranchCreationConfig_basicAuthCredentials - The basic authorization credentials for the autocreated branch.

$sel:buildSpec:AutoBranchCreationConfig', autoBranchCreationConfig_buildSpec - The build specification (build spec) for the autocreated branch.

$sel:enablePerformanceMode:AutoBranchCreationConfig', autoBranchCreationConfig_enablePerformanceMode - Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

$sel:environmentVariables:AutoBranchCreationConfig', autoBranchCreationConfig_environmentVariables - The environment variables for the autocreated branch.

$sel:enableAutoBuild:AutoBranchCreationConfig', autoBranchCreationConfig_enableAutoBuild - Enables auto building for the autocreated branch.

$sel:enableBasicAuth:AutoBranchCreationConfig', autoBranchCreationConfig_enableBasicAuth - Enables basic authorization for the autocreated branch.

$sel:pullRequestEnvironmentName:AutoBranchCreationConfig', autoBranchCreationConfig_pullRequestEnvironmentName - The Amplify environment name for the pull request.

autoBranchCreationConfig_framework :: Lens' AutoBranchCreationConfig (Maybe Text)

The framework for the autocreated branch.

autoBranchCreationConfig_stage :: Lens' AutoBranchCreationConfig (Maybe Stage)

Describes the current stage for the autocreated branch.

autoBranchCreationConfig_enablePullRequestPreview :: Lens' AutoBranchCreationConfig (Maybe Bool)

Enables pull request previews for the autocreated branch.

autoBranchCreationConfig_basicAuthCredentials :: Lens' AutoBranchCreationConfig (Maybe Text)

The basic authorization credentials for the autocreated branch.

autoBranchCreationConfig_buildSpec :: Lens' AutoBranchCreationConfig (Maybe Text)

The build specification (build spec) for the autocreated branch.

autoBranchCreationConfig_enablePerformanceMode :: Lens' AutoBranchCreationConfig (Maybe Bool)

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

autoBranchCreationConfig_environmentVariables :: Lens' AutoBranchCreationConfig (Maybe (HashMap Text Text))

The environment variables for the autocreated branch.

autoBranchCreationConfig_enableAutoBuild :: Lens' AutoBranchCreationConfig (Maybe Bool)

Enables auto building for the autocreated branch.

autoBranchCreationConfig_enableBasicAuth :: Lens' AutoBranchCreationConfig (Maybe Bool)

Enables basic authorization for the autocreated branch.

autoBranchCreationConfig_pullRequestEnvironmentName :: Lens' AutoBranchCreationConfig (Maybe Text)

The Amplify environment name for the pull request.

BackendEnvironment

data BackendEnvironment

Describes the backend environment for an Amplify app.

See: newBackendEnvironment smart constructor.

Constructors

BackendEnvironment'

Fields

deploymentArtifacts :: Maybe Text

The name of deployment artifacts.

stackName :: Maybe Text

The AWS CloudFormation stack name of a backend environment.

backendEnvironmentArn :: Text

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

environmentName :: Text

The name for a backend environment that is part of an Amplify app.

createTime :: POSIX

The creation date and time for a backend environment that is part of an Amplify app.

updateTime :: POSIX

The last updated date and time for a backend environment that is part of an Amplify app.

newBackendEnvironment

Arguments

:: Text
$sel:backendEnvironmentArn:BackendEnvironment'

-> Text
$sel:environmentName:BackendEnvironment'

-> UTCTime
$sel:createTime:BackendEnvironment'

-> UTCTime
$sel:updateTime:BackendEnvironment'

-> BackendEnvironment

Create a value of BackendEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentArtifacts:BackendEnvironment', backendEnvironment_deploymentArtifacts - The name of deployment artifacts.

$sel:stackName:BackendEnvironment', backendEnvironment_stackName - The AWS CloudFormation stack name of a backend environment.

$sel:backendEnvironmentArn:BackendEnvironment', backendEnvironment_backendEnvironmentArn - The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

$sel:environmentName:BackendEnvironment', backendEnvironment_environmentName - The name for a backend environment that is part of an Amplify app.

$sel:createTime:BackendEnvironment', backendEnvironment_createTime - The creation date and time for a backend environment that is part of an Amplify app.

$sel:updateTime:BackendEnvironment', backendEnvironment_updateTime - The last updated date and time for a backend environment that is part of an Amplify app.

backendEnvironment_deploymentArtifacts :: Lens' BackendEnvironment (Maybe Text)

The name of deployment artifacts.

backendEnvironment_stackName :: Lens' BackendEnvironment (Maybe Text)

The AWS CloudFormation stack name of a backend environment.

backendEnvironment_backendEnvironmentArn :: Lens' BackendEnvironment Text

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

backendEnvironment_environmentName :: Lens' BackendEnvironment Text

The name for a backend environment that is part of an Amplify app.

backendEnvironment_createTime :: Lens' BackendEnvironment UTCTime

The creation date and time for a backend environment that is part of an Amplify app.

backendEnvironment_updateTime :: Lens' BackendEnvironment UTCTime

The last updated date and time for a backend environment that is part of an Amplify app.

Branch

data Branch

The branch for an Amplify app, which maps to a third-party repository branch.

See: newBranch smart constructor.

Constructors

Branch'

Fields

backendEnvironmentArn :: Maybe Text

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

thumbnailUrl :: Maybe Text

The thumbnail URL for the branch of an Amplify app.

basicAuthCredentials :: Maybe (Sensitive Text)

The basic authorization credentials for a branch of an Amplify app.

buildSpec :: Maybe Text

The build specification (build spec) content for the branch of an Amplify app.

sourceBranch :: Maybe Text

The source branch if the branch is a pull request branch.

enablePerformanceMode :: Maybe Bool

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

destinationBranch :: Maybe Text

The destination branch if the branch is a pull request branch.

pullRequestEnvironmentName :: Maybe Text

The Amplify environment name for the pull request.

associatedResources :: Maybe [Text]

A list of custom resources that are linked to this branch.

tags :: Maybe (HashMap Text Text)

The tag for the branch of an Amplify app.

branchArn :: Text

The Amazon Resource Name (ARN) for a branch that is part of an Amplify app.

branchName :: Text

The name for the branch that is part of an Amplify app.

description :: Text

The description for the branch that is part of an Amplify app.

stage :: Stage

The current stage for the branch that is part of an Amplify app.

displayName :: Text

The display name for the branch. This is used as the default domain prefix.

enableNotification :: Bool

Enables notifications for a branch that is part of an Amplify app.

createTime :: POSIX

The creation date and time for a branch that is part of an Amplify app.

updateTime :: POSIX

The last updated date and time for a branch that is part of an Amplify app.

environmentVariables :: HashMap Text Text

The environment variables specific to a branch of an Amplify app.

enableAutoBuild :: Bool

Enables auto-building on push for a branch of an Amplify app.

customDomains :: [Text]

The custom domains for a branch of an Amplify app.

framework :: Text

The framework for a branch of an Amplify app.

activeJobId :: Text

The ID of the active job for a branch of an Amplify app.

totalNumberOfJobs :: Text

The total number of jobs that are part of an Amplify app.

enableBasicAuth :: Bool

Enables basic authorization for a branch of an Amplify app.

ttl :: Text

The content Time to Live (TTL) for the website in seconds.

enablePullRequestPreview :: Bool

Enables pull request previews for the branch.

newBranch

Arguments

:: Text
$sel:branchArn:Branch'

-> Text
$sel:branchName:Branch'

-> Text
$sel:description:Branch'

-> Stage
$sel:stage:Branch'

-> Text
$sel:displayName:Branch'

-> Bool
$sel:enableNotification:Branch'

-> UTCTime
$sel:createTime:Branch'

-> UTCTime
$sel:updateTime:Branch'

-> Bool
$sel:enableAutoBuild:Branch'

-> Text
$sel:framework:Branch'

-> Text
$sel:activeJobId:Branch'

-> Text
$sel:totalNumberOfJobs:Branch'

-> Bool
$sel:enableBasicAuth:Branch'

-> Text
$sel:ttl:Branch'

-> Bool
$sel:enablePullRequestPreview:Branch'

-> Branch

Create a value of Branch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backendEnvironmentArn:Branch', branch_backendEnvironmentArn - The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

$sel:thumbnailUrl:Branch', branch_thumbnailUrl - The thumbnail URL for the branch of an Amplify app.

$sel:basicAuthCredentials:Branch', branch_basicAuthCredentials - The basic authorization credentials for a branch of an Amplify app.

$sel:buildSpec:Branch', branch_buildSpec - The build specification (build spec) content for the branch of an Amplify app.

$sel:sourceBranch:Branch', branch_sourceBranch - The source branch if the branch is a pull request branch.

$sel:enablePerformanceMode:Branch', branch_enablePerformanceMode - Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

$sel:destinationBranch:Branch', branch_destinationBranch - The destination branch if the branch is a pull request branch.

$sel:pullRequestEnvironmentName:Branch', branch_pullRequestEnvironmentName - The Amplify environment name for the pull request.

$sel:associatedResources:Branch', branch_associatedResources - A list of custom resources that are linked to this branch.

$sel:tags:Branch', branch_tags - The tag for the branch of an Amplify app.

$sel:branchArn:Branch', branch_branchArn - The Amazon Resource Name (ARN) for a branch that is part of an Amplify app.

$sel:branchName:Branch', branch_branchName - The name for the branch that is part of an Amplify app.

$sel:description:Branch', branch_description - The description for the branch that is part of an Amplify app.

$sel:stage:Branch', branch_stage - The current stage for the branch that is part of an Amplify app.

$sel:displayName:Branch', branch_displayName - The display name for the branch. This is used as the default domain prefix.

$sel:enableNotification:Branch', branch_enableNotification - Enables notifications for a branch that is part of an Amplify app.

$sel:createTime:Branch', branch_createTime - The creation date and time for a branch that is part of an Amplify app.

$sel:updateTime:Branch', branch_updateTime - The last updated date and time for a branch that is part of an Amplify app.

$sel:environmentVariables:Branch', branch_environmentVariables - The environment variables specific to a branch of an Amplify app.

$sel:enableAutoBuild:Branch', branch_enableAutoBuild - Enables auto-building on push for a branch of an Amplify app.

$sel:customDomains:Branch', branch_customDomains - The custom domains for a branch of an Amplify app.

$sel:framework:Branch', branch_framework - The framework for a branch of an Amplify app.

$sel:activeJobId:Branch', branch_activeJobId - The ID of the active job for a branch of an Amplify app.

$sel:totalNumberOfJobs:Branch', branch_totalNumberOfJobs - The total number of jobs that are part of an Amplify app.

$sel:enableBasicAuth:Branch', branch_enableBasicAuth - Enables basic authorization for a branch of an Amplify app.

$sel:ttl:Branch', branch_ttl - The content Time to Live (TTL) for the website in seconds.

$sel:enablePullRequestPreview:Branch', branch_enablePullRequestPreview - Enables pull request previews for the branch.

branch_backendEnvironmentArn :: Lens' Branch (Maybe Text)

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

branch_thumbnailUrl :: Lens' Branch (Maybe Text)

The thumbnail URL for the branch of an Amplify app.

branch_basicAuthCredentials :: Lens' Branch (Maybe Text)

The basic authorization credentials for a branch of an Amplify app.

branch_buildSpec :: Lens' Branch (Maybe Text)

The build specification (build spec) content for the branch of an Amplify app.

branch_sourceBranch :: Lens' Branch (Maybe Text)

The source branch if the branch is a pull request branch.

branch_enablePerformanceMode :: Lens' Branch (Maybe Bool)

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

branch_destinationBranch :: Lens' Branch (Maybe Text)

The destination branch if the branch is a pull request branch.

branch_pullRequestEnvironmentName :: Lens' Branch (Maybe Text)

The Amplify environment name for the pull request.

branch_associatedResources :: Lens' Branch (Maybe [Text])

A list of custom resources that are linked to this branch.

branch_tags :: Lens' Branch (Maybe (HashMap Text Text))

The tag for the branch of an Amplify app.

branch_branchArn :: Lens' Branch Text

The Amazon Resource Name (ARN) for a branch that is part of an Amplify app.

branch_branchName :: Lens' Branch Text

The name for the branch that is part of an Amplify app.

branch_description :: Lens' Branch Text

The description for the branch that is part of an Amplify app.

branch_stage :: Lens' Branch Stage

The current stage for the branch that is part of an Amplify app.

branch_displayName :: Lens' Branch Text

The display name for the branch. This is used as the default domain prefix.

branch_enableNotification :: Lens' Branch Bool

Enables notifications for a branch that is part of an Amplify app.

branch_createTime :: Lens' Branch UTCTime

The creation date and time for a branch that is part of an Amplify app.

branch_updateTime :: Lens' Branch UTCTime

The last updated date and time for a branch that is part of an Amplify app.

branch_environmentVariables :: Lens' Branch (HashMap Text Text)

The environment variables specific to a branch of an Amplify app.

branch_enableAutoBuild :: Lens' Branch Bool

Enables auto-building on push for a branch of an Amplify app.

branch_customDomains :: Lens' Branch [Text]

The custom domains for a branch of an Amplify app.

branch_framework :: Lens' Branch Text

The framework for a branch of an Amplify app.

branch_activeJobId :: Lens' Branch Text

The ID of the active job for a branch of an Amplify app.

branch_totalNumberOfJobs :: Lens' Branch Text

The total number of jobs that are part of an Amplify app.

branch_enableBasicAuth :: Lens' Branch Bool

Enables basic authorization for a branch of an Amplify app.

branch_ttl :: Lens' Branch Text

The content Time to Live (TTL) for the website in seconds.

branch_enablePullRequestPreview :: Lens' Branch Bool

Enables pull request previews for the branch.

CustomRule

data CustomRule

Describes a custom rewrite or redirect rule.

See: newCustomRule smart constructor.

Constructors

CustomRule'

Fields

status :: Maybe Text

The status code for a URL rewrite or redirect rule.

200

Represents a 200 rewrite rule.

301

Represents a 301 (moved pemanently) redirect rule. This and all future requests should be directed to the target URL.

302

Represents a 302 temporary redirect rule.

404

Represents a 404 redirect rule.

404-200

Represents a 404 rewrite rule.

condition :: Maybe Text

The condition for a URL rewrite or redirect rule, such as a country code.

source :: Text

The source pattern for a URL rewrite or redirect rule.

target :: Text

The target pattern for a URL rewrite or redirect rule.

newCustomRule

Arguments

:: Text
$sel:source:CustomRule'

-> Text
$sel:target:CustomRule'

-> CustomRule

Create a value of CustomRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CustomRule', customRule_status - The status code for a URL rewrite or redirect rule.

200

Represents a 200 rewrite rule.

301

Represents a 301 (moved pemanently) redirect rule. This and all future requests should be directed to the target URL.

302

Represents a 302 temporary redirect rule.

404

Represents a 404 redirect rule.

404-200

Represents a 404 rewrite rule.

$sel:condition:CustomRule', customRule_condition - The condition for a URL rewrite or redirect rule, such as a country code.

$sel:source:CustomRule', customRule_source - The source pattern for a URL rewrite or redirect rule.

$sel:target:CustomRule', customRule_target - The target pattern for a URL rewrite or redirect rule.

customRule_status :: Lens' CustomRule (Maybe Text)

The status code for a URL rewrite or redirect rule.

200

Represents a 200 rewrite rule.

301

Represents a 301 (moved pemanently) redirect rule. This and all future requests should be directed to the target URL.

302

Represents a 302 temporary redirect rule.

404

Represents a 404 redirect rule.

404-200

Represents a 404 rewrite rule.

customRule_condition :: Lens' CustomRule (Maybe Text)

The condition for a URL rewrite or redirect rule, such as a country code.

customRule_source :: Lens' CustomRule Text

The source pattern for a URL rewrite or redirect rule.

customRule_target :: Lens' CustomRule Text

The target pattern for a URL rewrite or redirect rule.

DomainAssociation

data DomainAssociation

Describes a domain association that associates a custom domain with an Amplify app.

See: newDomainAssociation smart constructor.

Constructors

DomainAssociation'

Fields

certificateVerificationDNSRecord :: Maybe Text

The DNS record for certificate verification.

autoSubDomainCreationPatterns :: Maybe [Text]

Sets branch patterns for automatic subdomain creation.

autoSubDomainIAMRole :: Maybe Text

The required AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) for automatically creating subdomains.

domainAssociationArn :: Text

The Amazon Resource Name (ARN) for the domain association.

domainName :: Text

The name of the domain.

enableAutoSubDomain :: Bool

Enables the automated creation of subdomains for branches.

domainStatus :: DomainStatus

The current status of the domain association.

statusReason :: Text

The reason for the current status of the domain association.

subDomains :: [SubDomain]

The subdomains for the domain association.

newDomainAssociation

Arguments

:: Text
$sel:domainAssociationArn:DomainAssociation'

-> Text
$sel:domainName:DomainAssociation'

-> Bool
$sel:enableAutoSubDomain:DomainAssociation'

-> DomainStatus
$sel:domainStatus:DomainAssociation'

-> Text
$sel:statusReason:DomainAssociation'

-> DomainAssociation

Create a value of DomainAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateVerificationDNSRecord:DomainAssociation', domainAssociation_certificateVerificationDNSRecord - The DNS record for certificate verification.

$sel:autoSubDomainCreationPatterns:DomainAssociation', domainAssociation_autoSubDomainCreationPatterns - Sets branch patterns for automatic subdomain creation.

$sel:autoSubDomainIAMRole:DomainAssociation', domainAssociation_autoSubDomainIAMRole - The required AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) for automatically creating subdomains.

$sel:domainAssociationArn:DomainAssociation', domainAssociation_domainAssociationArn - The Amazon Resource Name (ARN) for the domain association.

$sel:domainName:DomainAssociation', domainAssociation_domainName - The name of the domain.

$sel:enableAutoSubDomain:DomainAssociation', domainAssociation_enableAutoSubDomain - Enables the automated creation of subdomains for branches.

$sel:domainStatus:DomainAssociation', domainAssociation_domainStatus - The current status of the domain association.

$sel:statusReason:DomainAssociation', domainAssociation_statusReason - The reason for the current status of the domain association.

$sel:subDomains:DomainAssociation', domainAssociation_subDomains - The subdomains for the domain association.

domainAssociation_certificateVerificationDNSRecord :: Lens' DomainAssociation (Maybe Text)

The DNS record for certificate verification.

domainAssociation_autoSubDomainCreationPatterns :: Lens' DomainAssociation (Maybe [Text])

Sets branch patterns for automatic subdomain creation.

domainAssociation_autoSubDomainIAMRole :: Lens' DomainAssociation (Maybe Text)

The required AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) for automatically creating subdomains.

domainAssociation_domainAssociationArn :: Lens' DomainAssociation Text

The Amazon Resource Name (ARN) for the domain association.

domainAssociation_domainName :: Lens' DomainAssociation Text

The name of the domain.

domainAssociation_enableAutoSubDomain :: Lens' DomainAssociation Bool

Enables the automated creation of subdomains for branches.

domainAssociation_domainStatus :: Lens' DomainAssociation DomainStatus

The current status of the domain association.

domainAssociation_statusReason :: Lens' DomainAssociation Text

The reason for the current status of the domain association.

domainAssociation_subDomains :: Lens' DomainAssociation [SubDomain]

The subdomains for the domain association.

Job

data Job

Describes an execution job for an Amplify app.

See: newJob smart constructor.

Constructors

Job'

Fields

summary :: JobSummary

Describes the summary for an execution job for an Amplify app.

steps :: [Step]

The execution steps for an execution job, for an Amplify app.

newJob

Arguments

:: JobSummary
$sel:summary:Job'

-> Job

Create a value of Job with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:Job', job_summary - Describes the summary for an execution job for an Amplify app.

$sel:steps:Job', job_steps - The execution steps for an execution job, for an Amplify app.

job_summary :: Lens' Job JobSummary

Describes the summary for an execution job for an Amplify app.

job_steps :: Lens' Job [Step]

The execution steps for an execution job, for an Amplify app.

JobSummary

data JobSummary

Describes the summary for an execution job for an Amplify app.

See: newJobSummary smart constructor.

Constructors

JobSummary'

Fields

endTime :: Maybe POSIX

The end date and time for the job.

jobArn :: Text

The Amazon Resource Name (ARN) for the job.

jobId :: Text

The unique ID for the job.

commitId :: Text

The commit ID from a third-party repository provider for the job.

commitMessage :: Text

The commit message from a third-party repository provider for the job.

commitTime :: POSIX

The commit date and time for the job.

startTime :: POSIX

The start date and time for the job.

status :: JobStatus

The current status for the job.

jobType :: JobType

The type for the job. If the value is RELEASE, the job was manually released from its source by using the StartJob API. If the value is RETRY, the job was manually retried using the StartJob API. If the value is WEB_HOOK, the job was automatically triggered by webhooks.

newJobSummary

Arguments

:: Text
$sel:jobArn:JobSummary'

-> Text
$sel:jobId:JobSummary'

-> Text
$sel:commitId:JobSummary'

-> Text
$sel:commitMessage:JobSummary'

-> UTCTime
$sel:commitTime:JobSummary'

-> UTCTime
$sel:startTime:JobSummary'

-> JobStatus
$sel:status:JobSummary'

-> JobType
$sel:jobType:JobSummary'

-> JobSummary

Create a value of JobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endTime:JobSummary', jobSummary_endTime - The end date and time for the job.

$sel:jobArn:JobSummary', jobSummary_jobArn - The Amazon Resource Name (ARN) for the job.

$sel:jobId:JobSummary', jobSummary_jobId - The unique ID for the job.

$sel:commitId:JobSummary', jobSummary_commitId - The commit ID from a third-party repository provider for the job.

$sel:commitMessage:JobSummary', jobSummary_commitMessage - The commit message from a third-party repository provider for the job.

$sel:commitTime:JobSummary', jobSummary_commitTime - The commit date and time for the job.

$sel:startTime:JobSummary', jobSummary_startTime - The start date and time for the job.

$sel:status:JobSummary', jobSummary_status - The current status for the job.

$sel:jobType:JobSummary', jobSummary_jobType - The type for the job. If the value is RELEASE, the job was manually released from its source by using the StartJob API. If the value is RETRY, the job was manually retried using the StartJob API. If the value is WEB_HOOK, the job was automatically triggered by webhooks.

jobSummary_endTime :: Lens' JobSummary (Maybe UTCTime)

The end date and time for the job.

jobSummary_jobArn :: Lens' JobSummary Text

The Amazon Resource Name (ARN) for the job.

jobSummary_jobId :: Lens' JobSummary Text

The unique ID for the job.

jobSummary_commitId :: Lens' JobSummary Text

The commit ID from a third-party repository provider for the job.

jobSummary_commitMessage :: Lens' JobSummary Text

The commit message from a third-party repository provider for the job.

jobSummary_commitTime :: Lens' JobSummary UTCTime

The commit date and time for the job.

jobSummary_startTime :: Lens' JobSummary UTCTime

The start date and time for the job.

jobSummary_status :: Lens' JobSummary JobStatus

The current status for the job.

jobSummary_jobType :: Lens' JobSummary JobType

The type for the job. If the value is RELEASE, the job was manually released from its source by using the StartJob API. If the value is RETRY, the job was manually retried using the StartJob API. If the value is WEB_HOOK, the job was automatically triggered by webhooks.

ProductionBranch

data ProductionBranch

Describes the information about a production branch for an Amplify app.

See: newProductionBranch smart constructor.

Constructors

ProductionBranch'

Fields

lastDeployTime :: Maybe POSIX

The last deploy time of the production branch.

status :: Maybe Text

The status of the production branch.

thumbnailUrl :: Maybe Text

The thumbnail URL for the production branch.

branchName :: Maybe Text

The branch name for the production branch.

newProductionBranch :: ProductionBranch

Create a value of ProductionBranch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastDeployTime:ProductionBranch', productionBranch_lastDeployTime - The last deploy time of the production branch.

$sel:status:ProductionBranch', productionBranch_status - The status of the production branch.

$sel:thumbnailUrl:ProductionBranch', productionBranch_thumbnailUrl - The thumbnail URL for the production branch.

$sel:branchName:ProductionBranch', productionBranch_branchName - The branch name for the production branch.

productionBranch_lastDeployTime :: Lens' ProductionBranch (Maybe UTCTime)

The last deploy time of the production branch.

productionBranch_status :: Lens' ProductionBranch (Maybe Text)

The status of the production branch.

productionBranch_thumbnailUrl :: Lens' ProductionBranch (Maybe Text)

The thumbnail URL for the production branch.

productionBranch_branchName :: Lens' ProductionBranch (Maybe Text)

The branch name for the production branch.

Step

data Step

Describes an execution step, for an execution job, for an Amplify app.

See: newStep smart constructor.

Constructors

Step'

Fields

logUrl :: Maybe Text

The URL to the logs for the execution step.

context :: Maybe Text

The context for the current step. Includes a build image if the step is build.

testArtifactsUrl :: Maybe Text

The URL to the test artifact for the execution step.

artifactsUrl :: Maybe Text

The URL to the artifact for the execution step.

testConfigUrl :: Maybe Text

The URL to the test configuration for the execution step.

screenshots :: Maybe (HashMap Text Text)

The list of screenshot URLs for the execution step, if relevant.

statusReason :: Maybe Text

The reason for the current step status.

stepName :: Text

The name of the execution step.

startTime :: POSIX

The start date and time of the execution step.

status :: JobStatus

The status of the execution step.

endTime :: POSIX

The end date and time of the execution step.

newStep

Arguments

:: Text
$sel:stepName:Step'

-> UTCTime
$sel:startTime:Step'

-> JobStatus
$sel:status:Step'

-> UTCTime
$sel:endTime:Step'

-> Step

Create a value of Step with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logUrl:Step', step_logUrl - The URL to the logs for the execution step.

$sel:context:Step', step_context - The context for the current step. Includes a build image if the step is build.

$sel:testArtifactsUrl:Step', step_testArtifactsUrl - The URL to the test artifact for the execution step.

$sel:artifactsUrl:Step', step_artifactsUrl - The URL to the artifact for the execution step.

$sel:testConfigUrl:Step', step_testConfigUrl - The URL to the test configuration for the execution step.

$sel:screenshots:Step', step_screenshots - The list of screenshot URLs for the execution step, if relevant.

$sel:statusReason:Step', step_statusReason - The reason for the current step status.

$sel:stepName:Step', step_stepName - The name of the execution step.

$sel:startTime:Step', step_startTime - The start date and time of the execution step.

$sel:status:Step', step_status - The status of the execution step.

$sel:endTime:Step', step_endTime - The end date and time of the execution step.

step_logUrl :: Lens' Step (Maybe Text)

The URL to the logs for the execution step.

step_context :: Lens' Step (Maybe Text)

The context for the current step. Includes a build image if the step is build.

step_testArtifactsUrl :: Lens' Step (Maybe Text)

The URL to the test artifact for the execution step.

step_artifactsUrl :: Lens' Step (Maybe Text)

The URL to the artifact for the execution step.

step_testConfigUrl :: Lens' Step (Maybe Text)

The URL to the test configuration for the execution step.

step_screenshots :: Lens' Step (Maybe (HashMap Text Text))

The list of screenshot URLs for the execution step, if relevant.

step_statusReason :: Lens' Step (Maybe Text)

The reason for the current step status.

step_stepName :: Lens' Step Text

The name of the execution step.

step_startTime :: Lens' Step UTCTime

The start date and time of the execution step.

step_status :: Lens' Step JobStatus

The status of the execution step.

step_endTime :: Lens' Step UTCTime

The end date and time of the execution step.

SubDomain

data SubDomain

The subdomain for the domain association.

See: newSubDomain smart constructor.

Constructors

SubDomain'

Fields

subDomainSetting :: SubDomainSetting

Describes the settings for the subdomain.

verified :: Bool

The verified status of the subdomain

dnsRecord :: Text

The DNS record for the subdomain.

newSubDomain

Arguments

:: SubDomainSetting
$sel:subDomainSetting:SubDomain'

-> Bool
$sel:verified:SubDomain'

-> Text
$sel:dnsRecord:SubDomain'

-> SubDomain

Create a value of SubDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subDomainSetting:SubDomain', subDomain_subDomainSetting - Describes the settings for the subdomain.

$sel:verified:SubDomain', subDomain_verified - The verified status of the subdomain

$sel:dnsRecord:SubDomain', subDomain_dnsRecord - The DNS record for the subdomain.

subDomain_subDomainSetting :: Lens' SubDomain SubDomainSetting

Describes the settings for the subdomain.

subDomain_verified :: Lens' SubDomain Bool

The verified status of the subdomain

subDomain_dnsRecord :: Lens' SubDomain Text

The DNS record for the subdomain.

SubDomainSetting

data SubDomainSetting

Describes the settings for the subdomain.

See: newSubDomainSetting smart constructor.

Constructors

SubDomainSetting'

Fields

prefix :: Text

The prefix setting for the subdomain.

branchName :: Text

The branch name setting for the subdomain.

newSubDomainSetting

Arguments

:: Text
$sel:prefix:SubDomainSetting'

-> Text
$sel:branchName:SubDomainSetting'

-> SubDomainSetting

Create a value of SubDomainSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:SubDomainSetting', subDomainSetting_prefix - The prefix setting for the subdomain.

$sel:branchName:SubDomainSetting', subDomainSetting_branchName - The branch name setting for the subdomain.

subDomainSetting_prefix :: Lens' SubDomainSetting Text

The prefix setting for the subdomain.

subDomainSetting_branchName :: Lens' SubDomainSetting Text

The branch name setting for the subdomain.

Webhook

data Webhook

Describes a webhook that connects repository events to an Amplify app.

See: newWebhook smart constructor.

Constructors

Webhook'

Fields

webhookArn :: Text

The Amazon Resource Name (ARN) for the webhook.

webhookId :: Text

The ID of the webhook.

webhookUrl :: Text

The URL of the webhook.

branchName :: Text

The name for a branch that is part of an Amplify app.

description :: Text

The description for a webhook.

createTime :: POSIX

The create date and time for a webhook.

updateTime :: POSIX

Updates the date and time for a webhook.

newWebhook

Arguments

:: Text
$sel:webhookArn:Webhook'

-> Text
$sel:webhookId:Webhook'

-> Text
$sel:webhookUrl:Webhook'

-> Text
$sel:branchName:Webhook'

-> Text
$sel:description:Webhook'

-> UTCTime
$sel:createTime:Webhook'

-> UTCTime
$sel:updateTime:Webhook'

-> Webhook

Create a value of Webhook with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:webhookArn:Webhook', webhook_webhookArn - The Amazon Resource Name (ARN) for the webhook.

$sel:webhookId:Webhook', webhook_webhookId - The ID of the webhook.

$sel:webhookUrl:Webhook', webhook_webhookUrl - The URL of the webhook.

$sel:branchName:Webhook', webhook_branchName - The name for a branch that is part of an Amplify app.

$sel:description:Webhook', webhook_description - The description for a webhook.

$sel:createTime:Webhook', webhook_createTime - The create date and time for a webhook.

$sel:updateTime:Webhook', webhook_updateTime - Updates the date and time for a webhook.

webhook_webhookArn :: Lens' Webhook Text

The Amazon Resource Name (ARN) for the webhook.

webhook_webhookId :: Lens' Webhook Text

The ID of the webhook.

webhook_webhookUrl :: Lens' Webhook Text

The URL of the webhook.

webhook_branchName :: Lens' Webhook Text

The name for a branch that is part of an Amplify app.

webhook_description :: Lens' Webhook Text

The description for a webhook.

webhook_createTime :: Lens' Webhook UTCTime

The create date and time for a webhook.

webhook_updateTime :: Lens' Webhook UTCTime

Updates the date and time for a webhook.