Network.AWS.Amplify.Types.Branch

Description

Documentation

data Branch

The branch for an Amplify app, which maps to a third-party repository branch.

See: newBranch smart constructor.

Constructors

Branch'

Fields

backendEnvironmentArn :: Maybe Text

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

thumbnailUrl :: Maybe Text

The thumbnail URL for the branch of an Amplify app.

basicAuthCredentials :: Maybe (Sensitive Text)

The basic authorization credentials for a branch of an Amplify app.

buildSpec :: Maybe Text

The build specification (build spec) content for the branch of an Amplify app.

sourceBranch :: Maybe Text

The source branch if the branch is a pull request branch.

enablePerformanceMode :: Maybe Bool

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

destinationBranch :: Maybe Text

The destination branch if the branch is a pull request branch.

pullRequestEnvironmentName :: Maybe Text

The Amplify environment name for the pull request.

associatedResources :: Maybe [Text]

A list of custom resources that are linked to this branch.

tags :: Maybe (HashMap Text Text)

The tag for the branch of an Amplify app.

branchArn :: Text

The Amazon Resource Name (ARN) for a branch that is part of an Amplify app.

branchName :: Text

The name for the branch that is part of an Amplify app.

description :: Text

The description for the branch that is part of an Amplify app.

stage :: Stage

The current stage for the branch that is part of an Amplify app.

displayName :: Text

The display name for the branch. This is used as the default domain prefix.

enableNotification :: Bool

Enables notifications for a branch that is part of an Amplify app.

createTime :: POSIX

The creation date and time for a branch that is part of an Amplify app.

updateTime :: POSIX

The last updated date and time for a branch that is part of an Amplify app.

environmentVariables :: HashMap Text Text

The environment variables specific to a branch of an Amplify app.

enableAutoBuild :: Bool

Enables auto-building on push for a branch of an Amplify app.

customDomains :: [Text]

The custom domains for a branch of an Amplify app.

framework :: Text

The framework for a branch of an Amplify app.

activeJobId :: Text

The ID of the active job for a branch of an Amplify app.

totalNumberOfJobs :: Text

The total number of jobs that are part of an Amplify app.

enableBasicAuth :: Bool

Enables basic authorization for a branch of an Amplify app.

ttl :: Text

The content Time to Live (TTL) for the website in seconds.

enablePullRequestPreview :: Bool

Enables pull request previews for the branch.

newBranch

Arguments

:: Text
$sel:branchArn:Branch'

-> Text
$sel:branchName:Branch'

-> Text
$sel:description:Branch'

-> Stage
$sel:stage:Branch'

-> Text
$sel:displayName:Branch'

-> Bool
$sel:enableNotification:Branch'

-> UTCTime
$sel:createTime:Branch'

-> UTCTime
$sel:updateTime:Branch'

-> Bool
$sel:enableAutoBuild:Branch'

-> Text
$sel:framework:Branch'

-> Text
$sel:activeJobId:Branch'

-> Text
$sel:totalNumberOfJobs:Branch'

-> Bool
$sel:enableBasicAuth:Branch'

-> Text
$sel:ttl:Branch'

-> Bool
$sel:enablePullRequestPreview:Branch'

-> Branch

Create a value of Branch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backendEnvironmentArn:Branch', branch_backendEnvironmentArn - The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

$sel:thumbnailUrl:Branch', branch_thumbnailUrl - The thumbnail URL for the branch of an Amplify app.

$sel:basicAuthCredentials:Branch', branch_basicAuthCredentials - The basic authorization credentials for a branch of an Amplify app.

$sel:buildSpec:Branch', branch_buildSpec - The build specification (build spec) content for the branch of an Amplify app.

$sel:sourceBranch:Branch', branch_sourceBranch - The source branch if the branch is a pull request branch.

$sel:enablePerformanceMode:Branch', branch_enablePerformanceMode - Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

$sel:destinationBranch:Branch', branch_destinationBranch - The destination branch if the branch is a pull request branch.

$sel:pullRequestEnvironmentName:Branch', branch_pullRequestEnvironmentName - The Amplify environment name for the pull request.

$sel:associatedResources:Branch', branch_associatedResources - A list of custom resources that are linked to this branch.

$sel:tags:Branch', branch_tags - The tag for the branch of an Amplify app.

$sel:branchArn:Branch', branch_branchArn - The Amazon Resource Name (ARN) for a branch that is part of an Amplify app.

$sel:branchName:Branch', branch_branchName - The name for the branch that is part of an Amplify app.

$sel:description:Branch', branch_description - The description for the branch that is part of an Amplify app.

$sel:stage:Branch', branch_stage - The current stage for the branch that is part of an Amplify app.

$sel:displayName:Branch', branch_displayName - The display name for the branch. This is used as the default domain prefix.

$sel:enableNotification:Branch', branch_enableNotification - Enables notifications for a branch that is part of an Amplify app.

$sel:createTime:Branch', branch_createTime - The creation date and time for a branch that is part of an Amplify app.

$sel:updateTime:Branch', branch_updateTime - The last updated date and time for a branch that is part of an Amplify app.

$sel:environmentVariables:Branch', branch_environmentVariables - The environment variables specific to a branch of an Amplify app.

$sel:enableAutoBuild:Branch', branch_enableAutoBuild - Enables auto-building on push for a branch of an Amplify app.

$sel:customDomains:Branch', branch_customDomains - The custom domains for a branch of an Amplify app.

$sel:framework:Branch', branch_framework - The framework for a branch of an Amplify app.

$sel:activeJobId:Branch', branch_activeJobId - The ID of the active job for a branch of an Amplify app.

$sel:totalNumberOfJobs:Branch', branch_totalNumberOfJobs - The total number of jobs that are part of an Amplify app.

$sel:enableBasicAuth:Branch', branch_enableBasicAuth - Enables basic authorization for a branch of an Amplify app.

$sel:ttl:Branch', branch_ttl - The content Time to Live (TTL) for the website in seconds.

$sel:enablePullRequestPreview:Branch', branch_enablePullRequestPreview - Enables pull request previews for the branch.

branch_backendEnvironmentArn :: Lens' Branch (Maybe Text)

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

branch_thumbnailUrl :: Lens' Branch (Maybe Text)

The thumbnail URL for the branch of an Amplify app.

branch_basicAuthCredentials :: Lens' Branch (Maybe Text)

The basic authorization credentials for a branch of an Amplify app.

branch_buildSpec :: Lens' Branch (Maybe Text)

The build specification (build spec) content for the branch of an Amplify app.

branch_sourceBranch :: Lens' Branch (Maybe Text)

The source branch if the branch is a pull request branch.

branch_enablePerformanceMode :: Lens' Branch (Maybe Bool)

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

branch_destinationBranch :: Lens' Branch (Maybe Text)

The destination branch if the branch is a pull request branch.

branch_pullRequestEnvironmentName :: Lens' Branch (Maybe Text)

The Amplify environment name for the pull request.

branch_associatedResources :: Lens' Branch (Maybe [Text])

A list of custom resources that are linked to this branch.

branch_tags :: Lens' Branch (Maybe (HashMap Text Text))

The tag for the branch of an Amplify app.

branch_branchArn :: Lens' Branch Text

The Amazon Resource Name (ARN) for a branch that is part of an Amplify app.

branch_branchName :: Lens' Branch Text

The name for the branch that is part of an Amplify app.

branch_description :: Lens' Branch Text

The description for the branch that is part of an Amplify app.

branch_stage :: Lens' Branch Stage

The current stage for the branch that is part of an Amplify app.

branch_displayName :: Lens' Branch Text

The display name for the branch. This is used as the default domain prefix.

branch_enableNotification :: Lens' Branch Bool

Enables notifications for a branch that is part of an Amplify app.

branch_createTime :: Lens' Branch UTCTime

The creation date and time for a branch that is part of an Amplify app.

branch_updateTime :: Lens' Branch UTCTime

The last updated date and time for a branch that is part of an Amplify app.

branch_environmentVariables :: Lens' Branch (HashMap Text Text)

The environment variables specific to a branch of an Amplify app.

branch_enableAutoBuild :: Lens' Branch Bool

Enables auto-building on push for a branch of an Amplify app.

branch_customDomains :: Lens' Branch [Text]

The custom domains for a branch of an Amplify app.

branch_framework :: Lens' Branch Text

The framework for a branch of an Amplify app.

branch_activeJobId :: Lens' Branch Text

The ID of the active job for a branch of an Amplify app.

branch_totalNumberOfJobs :: Lens' Branch Text

The total number of jobs that are part of an Amplify app.

branch_enableBasicAuth :: Lens' Branch Bool

Enables basic authorization for a branch of an Amplify app.

branch_ttl :: Lens' Branch Text

The content Time to Live (TTL) for the website in seconds.

branch_enablePullRequestPreview :: Lens' Branch Bool

Enables pull request previews for the branch.