Network.AWS.Amplify.UpdateApp

Description

Updates an existing Amplify app.

Creating a Request

data UpdateApp

The request structure for the update app request.

See: newUpdateApp smart constructor.

Constructors

UpdateApp'

Fields

enableBranchAutoBuild :: Maybe Bool

Enables branch auto-building for an Amplify app.

oauthToken :: Maybe (Sensitive Text)

The OAuth token for a third-party source control system for an Amplify app. The token is used to create a webhook and a read-only deploy key. The OAuth token is not stored.

accessToken :: Maybe (Sensitive Text)

The personal access token for a third-party source control system for an Amplify app. The token is used to create webhook and a read-only deploy key. The token is not stored.

customHeaders :: Maybe Text

The custom HTTP headers for an Amplify app.

platform :: Maybe Platform

The platform for an Amplify app.

basicAuthCredentials :: Maybe (Sensitive Text)

The basic authorization credentials for an Amplify app.

repository :: Maybe Text

The name of the repository for an Amplify app

buildSpec :: Maybe Text

The build specification (build spec) for an Amplify app.

enableBranchAutoDeletion :: Maybe Bool

Automatically disconnects a branch in the Amplify Console when you delete a branch from your Git repository.

customRules :: Maybe [CustomRule]

The custom redirect and rewrite rules for an Amplify app.

iamServiceRoleArn :: Maybe Text

The AWS Identity and Access Management (IAM) service role for an Amplify app.

autoBranchCreationPatterns :: Maybe [Text]

Describes the automated branch creation glob patterns for an Amplify app.

name :: Maybe Text

The name for an Amplify app.

autoBranchCreationConfig :: Maybe AutoBranchCreationConfig

The automated branch creation configuration for an Amplify app.

environmentVariables :: Maybe (HashMap Text Text)

The environment variables for an Amplify app.

enableAutoBranchCreation :: Maybe Bool

Enables automated branch creation for an Amplify app.

enableBasicAuth :: Maybe Bool

Enables basic authorization for an Amplify app.

description :: Maybe Text

The description for an Amplify app.

appId :: Text

The unique ID for an Amplify app.

newUpdateApp

Arguments

:: Text
$sel:appId:UpdateApp'

-> UpdateApp

Create a value of UpdateApp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enableBranchAutoBuild:UpdateApp', updateApp_enableBranchAutoBuild - Enables branch auto-building for an Amplify app.

$sel:oauthToken:UpdateApp', updateApp_oauthToken - The OAuth token for a third-party source control system for an Amplify app. The token is used to create a webhook and a read-only deploy key. The OAuth token is not stored.

$sel:accessToken:UpdateApp', updateApp_accessToken - The personal access token for a third-party source control system for an Amplify app. The token is used to create webhook and a read-only deploy key. The token is not stored.

$sel:customHeaders:UpdateApp', updateApp_customHeaders - The custom HTTP headers for an Amplify app.

$sel:platform:UpdateApp', updateApp_platform - The platform for an Amplify app.

$sel:basicAuthCredentials:UpdateApp', updateApp_basicAuthCredentials - The basic authorization credentials for an Amplify app.

$sel:repository:UpdateApp', updateApp_repository - The name of the repository for an Amplify app

$sel:buildSpec:UpdateApp', updateApp_buildSpec - The build specification (build spec) for an Amplify app.

$sel:enableBranchAutoDeletion:UpdateApp', updateApp_enableBranchAutoDeletion - Automatically disconnects a branch in the Amplify Console when you delete a branch from your Git repository.

$sel:customRules:UpdateApp', updateApp_customRules - The custom redirect and rewrite rules for an Amplify app.

$sel:iamServiceRoleArn:UpdateApp', updateApp_iamServiceRoleArn - The AWS Identity and Access Management (IAM) service role for an Amplify app.

$sel:autoBranchCreationPatterns:UpdateApp', updateApp_autoBranchCreationPatterns - Describes the automated branch creation glob patterns for an Amplify app.

$sel:name:UpdateApp', updateApp_name - The name for an Amplify app.

$sel:autoBranchCreationConfig:UpdateApp', updateApp_autoBranchCreationConfig - The automated branch creation configuration for an Amplify app.

$sel:environmentVariables:UpdateApp', updateApp_environmentVariables - The environment variables for an Amplify app.

$sel:enableAutoBranchCreation:UpdateApp', updateApp_enableAutoBranchCreation - Enables automated branch creation for an Amplify app.

$sel:enableBasicAuth:UpdateApp', updateApp_enableBasicAuth - Enables basic authorization for an Amplify app.

$sel:description:UpdateApp', updateApp_description - The description for an Amplify app.

$sel:appId:UpdateApp', updateApp_appId - The unique ID for an Amplify app.

Request Lenses

updateApp_enableBranchAutoBuild :: Lens' UpdateApp (Maybe Bool)

Enables branch auto-building for an Amplify app.

updateApp_oauthToken :: Lens' UpdateApp (Maybe Text)

The OAuth token for a third-party source control system for an Amplify app. The token is used to create a webhook and a read-only deploy key. The OAuth token is not stored.

updateApp_accessToken :: Lens' UpdateApp (Maybe Text)

The personal access token for a third-party source control system for an Amplify app. The token is used to create webhook and a read-only deploy key. The token is not stored.

updateApp_customHeaders :: Lens' UpdateApp (Maybe Text)

The custom HTTP headers for an Amplify app.

updateApp_platform :: Lens' UpdateApp (Maybe Platform)

The platform for an Amplify app.

updateApp_basicAuthCredentials :: Lens' UpdateApp (Maybe Text)

The basic authorization credentials for an Amplify app.

updateApp_repository :: Lens' UpdateApp (Maybe Text)

The name of the repository for an Amplify app

updateApp_buildSpec :: Lens' UpdateApp (Maybe Text)

The build specification (build spec) for an Amplify app.

updateApp_enableBranchAutoDeletion :: Lens' UpdateApp (Maybe Bool)

Automatically disconnects a branch in the Amplify Console when you delete a branch from your Git repository.

updateApp_customRules :: Lens' UpdateApp (Maybe [CustomRule])

The custom redirect and rewrite rules for an Amplify app.

updateApp_iamServiceRoleArn :: Lens' UpdateApp (Maybe Text)

The AWS Identity and Access Management (IAM) service role for an Amplify app.

updateApp_autoBranchCreationPatterns :: Lens' UpdateApp (Maybe [Text])

Describes the automated branch creation glob patterns for an Amplify app.

updateApp_name :: Lens' UpdateApp (Maybe Text)

The name for an Amplify app.

updateApp_autoBranchCreationConfig :: Lens' UpdateApp (Maybe AutoBranchCreationConfig)

The automated branch creation configuration for an Amplify app.

updateApp_environmentVariables :: Lens' UpdateApp (Maybe (HashMap Text Text))

The environment variables for an Amplify app.

updateApp_enableAutoBranchCreation :: Lens' UpdateApp (Maybe Bool)

Enables automated branch creation for an Amplify app.

updateApp_enableBasicAuth :: Lens' UpdateApp (Maybe Bool)

Enables basic authorization for an Amplify app.

updateApp_description :: Lens' UpdateApp (Maybe Text)

The description for an Amplify app.

updateApp_appId :: Lens' UpdateApp Text

The unique ID for an Amplify app.

Destructuring the Response

data UpdateAppResponse

The result structure for an Amplify app update request.

See: newUpdateAppResponse smart constructor.

Constructors

UpdateAppResponse'

Fields

httpStatus :: Int

The response's http status code.

app :: App

Represents the updated Amplify app.

newUpdateAppResponse

Arguments

:: Int
$sel:httpStatus:UpdateAppResponse'

-> App
$sel:app:UpdateAppResponse'

-> UpdateAppResponse

Create a value of UpdateAppResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateAppResponse', updateAppResponse_httpStatus - The response's http status code.

$sel:app:UpdateAppResponse', updateAppResponse_app - Represents the updated Amplify app.

Response Lenses

updateAppResponse_httpStatus :: Lens' UpdateAppResponse Int

The response's http status code.

updateAppResponse_app :: Lens' UpdateAppResponse App

Represents the updated Amplify app.