Network.AWS.Amplify.UpdateBranch

Description

Updates a branch for an Amplify app.

Creating a Request

data UpdateBranch

The request structure for the update branch request.

See: newUpdateBranch smart constructor.

Constructors

UpdateBranch'

Fields

framework :: Maybe Text

The framework for the branch.

ttl :: Maybe Text

The content Time to Live (TTL) for the website in seconds.

enableNotification :: Maybe Bool

Enables notifications for the branch.

stage :: Maybe Stage

Describes the current stage for the branch.

backendEnvironmentArn :: Maybe Text

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

enablePullRequestPreview :: Maybe Bool

Enables pull request previews for this branch.

basicAuthCredentials :: Maybe (Sensitive Text)

The basic authorization credentials for the branch.

buildSpec :: Maybe Text

The build specification (build spec) for the branch.

enablePerformanceMode :: Maybe Bool

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

displayName :: Maybe Text

The display name for a branch. This is used as the default domain prefix.

environmentVariables :: Maybe (HashMap Text Text)

The environment variables for the branch.

enableAutoBuild :: Maybe Bool

Enables auto building for the branch.

enableBasicAuth :: Maybe Bool

Enables basic authorization for the branch.

pullRequestEnvironmentName :: Maybe Text

The Amplify environment name for the pull request.

description :: Maybe Text

The description for the branch.

appId :: Text

The unique ID for an Amplify app.

branchName :: Text

The name for the branch.

newUpdateBranch

Arguments

:: Text
$sel:appId:UpdateBranch'

-> Text
$sel:branchName:UpdateBranch'

-> UpdateBranch

Create a value of UpdateBranch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:framework:UpdateBranch', updateBranch_framework - The framework for the branch.

$sel:ttl:UpdateBranch', updateBranch_ttl - The content Time to Live (TTL) for the website in seconds.

$sel:enableNotification:UpdateBranch', updateBranch_enableNotification - Enables notifications for the branch.

$sel:stage:UpdateBranch', updateBranch_stage - Describes the current stage for the branch.

$sel:backendEnvironmentArn:UpdateBranch', updateBranch_backendEnvironmentArn - The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

$sel:enablePullRequestPreview:UpdateBranch', updateBranch_enablePullRequestPreview - Enables pull request previews for this branch.

$sel:basicAuthCredentials:UpdateBranch', updateBranch_basicAuthCredentials - The basic authorization credentials for the branch.

$sel:buildSpec:UpdateBranch', updateBranch_buildSpec - The build specification (build spec) for the branch.

$sel:enablePerformanceMode:UpdateBranch', updateBranch_enablePerformanceMode - Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

$sel:displayName:UpdateBranch', updateBranch_displayName - The display name for a branch. This is used as the default domain prefix.

$sel:environmentVariables:UpdateBranch', updateBranch_environmentVariables - The environment variables for the branch.

$sel:enableAutoBuild:UpdateBranch', updateBranch_enableAutoBuild - Enables auto building for the branch.

$sel:enableBasicAuth:UpdateBranch', updateBranch_enableBasicAuth - Enables basic authorization for the branch.

$sel:pullRequestEnvironmentName:UpdateBranch', updateBranch_pullRequestEnvironmentName - The Amplify environment name for the pull request.

$sel:description:UpdateBranch', updateBranch_description - The description for the branch.

$sel:appId:UpdateBranch', updateBranch_appId - The unique ID for an Amplify app.

$sel:branchName:UpdateBranch', updateBranch_branchName - The name for the branch.

Request Lenses

updateBranch_framework :: Lens' UpdateBranch (Maybe Text)

The framework for the branch.

updateBranch_ttl :: Lens' UpdateBranch (Maybe Text)

The content Time to Live (TTL) for the website in seconds.

updateBranch_enableNotification :: Lens' UpdateBranch (Maybe Bool)

Enables notifications for the branch.

updateBranch_stage :: Lens' UpdateBranch (Maybe Stage)

Describes the current stage for the branch.

updateBranch_backendEnvironmentArn :: Lens' UpdateBranch (Maybe Text)

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

updateBranch_enablePullRequestPreview :: Lens' UpdateBranch (Maybe Bool)

Enables pull request previews for this branch.

updateBranch_basicAuthCredentials :: Lens' UpdateBranch (Maybe Text)

The basic authorization credentials for the branch.

updateBranch_buildSpec :: Lens' UpdateBranch (Maybe Text)

The build specification (build spec) for the branch.

updateBranch_enablePerformanceMode :: Lens' UpdateBranch (Maybe Bool)

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

updateBranch_displayName :: Lens' UpdateBranch (Maybe Text)

The display name for a branch. This is used as the default domain prefix.

updateBranch_environmentVariables :: Lens' UpdateBranch (Maybe (HashMap Text Text))

The environment variables for the branch.

updateBranch_enableAutoBuild :: Lens' UpdateBranch (Maybe Bool)

Enables auto building for the branch.

updateBranch_enableBasicAuth :: Lens' UpdateBranch (Maybe Bool)

Enables basic authorization for the branch.

updateBranch_pullRequestEnvironmentName :: Lens' UpdateBranch (Maybe Text)

The Amplify environment name for the pull request.

updateBranch_description :: Lens' UpdateBranch (Maybe Text)

The description for the branch.

updateBranch_appId :: Lens' UpdateBranch Text

The unique ID for an Amplify app.

updateBranch_branchName :: Lens' UpdateBranch Text

The name for the branch.

Destructuring the Response

data UpdateBranchResponse

The result structure for the update branch request.

See: newUpdateBranchResponse smart constructor.

Constructors

UpdateBranchResponse'

Fields

httpStatus :: Int

The response's http status code.

branch :: Branch

The branch for an Amplify app, which maps to a third-party repository branch.

newUpdateBranchResponse

Arguments

:: Int
$sel:httpStatus:UpdateBranchResponse'

-> Branch
$sel:branch:UpdateBranchResponse'

-> UpdateBranchResponse

Create a value of UpdateBranchResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateBranchResponse', updateBranchResponse_httpStatus - The response's http status code.

$sel:branch:UpdateBranchResponse', updateBranchResponse_branch - The branch for an Amplify app, which maps to a third-party repository branch.

Response Lenses

updateBranchResponse_httpStatus :: Lens' UpdateBranchResponse Int

The response's http status code.

updateBranchResponse_branch :: Lens' UpdateBranchResponse Branch

The branch for an Amplify app, which maps to a third-party repository branch.