Network.AWS.Amplify.UpdateWebhook

Description

Updates a webhook.

Creating a Request

data UpdateWebhook

The request structure for the update webhook request.

See: newUpdateWebhook smart constructor.

Constructors

UpdateWebhook'

Fields

branchName :: Maybe Text

The name for a branch that is part of an Amplify app.

description :: Maybe Text

The description for a webhook.

webhookId :: Text

The unique ID for a webhook.

newUpdateWebhook

Arguments

:: Text
$sel:webhookId:UpdateWebhook'

-> UpdateWebhook

Create a value of UpdateWebhook with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:branchName:UpdateWebhook', updateWebhook_branchName - The name for a branch that is part of an Amplify app.

$sel:description:UpdateWebhook', updateWebhook_description - The description for a webhook.

$sel:webhookId:UpdateWebhook', updateWebhook_webhookId - The unique ID for a webhook.

Request Lenses

updateWebhook_branchName :: Lens' UpdateWebhook (Maybe Text)

The name for a branch that is part of an Amplify app.

updateWebhook_description :: Lens' UpdateWebhook (Maybe Text)

The description for a webhook.

updateWebhook_webhookId :: Lens' UpdateWebhook Text

The unique ID for a webhook.

Destructuring the Response

data UpdateWebhookResponse

The result structure for the update webhook request.

See: newUpdateWebhookResponse smart constructor.

Constructors

UpdateWebhookResponse'

Fields

httpStatus :: Int

The response's http status code.

webhook :: Webhook

Describes a webhook that connects repository events to an Amplify app.

newUpdateWebhookResponse

Arguments

:: Int
$sel:httpStatus:UpdateWebhookResponse'

-> Webhook
$sel:webhook:UpdateWebhookResponse'

-> UpdateWebhookResponse

Create a value of UpdateWebhookResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateWebhookResponse', updateWebhookResponse_httpStatus - The response's http status code.

$sel:webhook:UpdateWebhookResponse', updateWebhookResponse_webhook - Describes a webhook that connects repository events to an Amplify app.

Response Lenses

updateWebhookResponse_httpStatus :: Lens' UpdateWebhookResponse Int

The response's http status code.

updateWebhookResponse_webhook :: Lens' UpdateWebhookResponse Webhook

Describes a webhook that connects repository events to an Amplify app.