Network.AWS.AmplifyBackend.CreateBackendAuth

Description

Creates a new backend authentication resource.

Creating a Request

data CreateBackendAuth

The request body for CreateBackendAuth.

See: newCreateBackendAuth smart constructor.

Constructors

CreateBackendAuth'

Fields

appId :: Text

The app ID.

resourceName :: Text

The name of this resource.

backendEnvironmentName :: Text

The name of the backend environment.

resourceConfig :: CreateBackendAuthResourceConfig

The resource configuration for this request object.

newCreateBackendAuth

Arguments

:: Text
$sel:appId:CreateBackendAuth'

-> Text
$sel:resourceName:CreateBackendAuth'

-> Text
$sel:backendEnvironmentName:CreateBackendAuth'

-> CreateBackendAuthResourceConfig
$sel:resourceConfig:CreateBackendAuth'

-> CreateBackendAuth

Create a value of CreateBackendAuth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:CreateBackendAuth', createBackendAuth_appId - The app ID.

$sel:resourceName:CreateBackendAuth', createBackendAuth_resourceName - The name of this resource.

$sel:backendEnvironmentName:CreateBackendAuth', createBackendAuth_backendEnvironmentName - The name of the backend environment.

$sel:resourceConfig:CreateBackendAuth', createBackendAuth_resourceConfig - The resource configuration for this request object.

Request Lenses

createBackendAuth_appId :: Lens' CreateBackendAuth Text

The app ID.

createBackendAuth_resourceName :: Lens' CreateBackendAuth Text

The name of this resource.

createBackendAuth_backendEnvironmentName :: Lens' CreateBackendAuth Text

The name of the backend environment.

createBackendAuth_resourceConfig :: Lens' CreateBackendAuth CreateBackendAuthResourceConfig

The resource configuration for this request object.

Destructuring the Response

data CreateBackendAuthResponse

See: newCreateBackendAuthResponse smart constructor.

Constructors

CreateBackendAuthResponse'

Fields

status :: Maybe Text

The current status of the request.

jobId :: Maybe Text

The ID for the job.

operation :: Maybe Text

The name of the operation.

error :: Maybe Text

If the request fails, this error is returned.

appId :: Maybe Text

The app ID.

backendEnvironmentName :: Maybe Text

The name of the backend environment.

httpStatus :: Int

The response's http status code.

newCreateBackendAuthResponse

Arguments

:: Int
$sel:httpStatus:CreateBackendAuthResponse'

-> CreateBackendAuthResponse

Create a value of CreateBackendAuthResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateBackendAuthResponse', createBackendAuthResponse_status - The current status of the request.

$sel:jobId:CreateBackendAuthResponse', createBackendAuthResponse_jobId - The ID for the job.

$sel:operation:CreateBackendAuthResponse', createBackendAuthResponse_operation - The name of the operation.

$sel:error:CreateBackendAuthResponse', createBackendAuthResponse_error - If the request fails, this error is returned.

$sel:appId:CreateBackendAuth', createBackendAuthResponse_appId - The app ID.

$sel:backendEnvironmentName:CreateBackendAuth', createBackendAuthResponse_backendEnvironmentName - The name of the backend environment.

$sel:httpStatus:CreateBackendAuthResponse', createBackendAuthResponse_httpStatus - The response's http status code.

Response Lenses

createBackendAuthResponse_status :: Lens' CreateBackendAuthResponse (Maybe Text)

The current status of the request.

createBackendAuthResponse_jobId :: Lens' CreateBackendAuthResponse (Maybe Text)

The ID for the job.

createBackendAuthResponse_operation :: Lens' CreateBackendAuthResponse (Maybe Text)

The name of the operation.

createBackendAuthResponse_error :: Lens' CreateBackendAuthResponse (Maybe Text)

If the request fails, this error is returned.

createBackendAuthResponse_appId :: Lens' CreateBackendAuthResponse (Maybe Text)

The app ID.

createBackendAuthResponse_backendEnvironmentName :: Lens' CreateBackendAuthResponse (Maybe Text)

The name of the backend environment.

createBackendAuthResponse_httpStatus :: Lens' CreateBackendAuthResponse Int

The response's http status code.