Network.AWS.AmplifyBackend.DeleteBackend

Description

Removes an existing environment from your Amplify project.

Creating a Request

data DeleteBackend

See: newDeleteBackend smart constructor.

Constructors

DeleteBackend'

Fields

appId :: Text

The app ID.

backendEnvironmentName :: Text

The name of the backend environment.

newDeleteBackend

Arguments

:: Text
$sel:appId:DeleteBackend'

-> Text
$sel:backendEnvironmentName:DeleteBackend'

-> DeleteBackend

Create a value of DeleteBackend with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:DeleteBackend', deleteBackend_appId - The app ID.

$sel:backendEnvironmentName:DeleteBackend', deleteBackend_backendEnvironmentName - The name of the backend environment.

Request Lenses

deleteBackend_appId :: Lens' DeleteBackend Text

The app ID.

deleteBackend_backendEnvironmentName :: Lens' DeleteBackend Text

The name of the backend environment.

Destructuring the Response

data DeleteBackendResponse

See: newDeleteBackendResponse smart constructor.

Constructors

DeleteBackendResponse'

Fields

status :: Maybe Text

The current status of the request.

jobId :: Maybe Text

The ID for the job.

operation :: Maybe Text

The name of the operation.

error :: Maybe Text

If the request fails, this error is returned.

appId :: Maybe Text

The app ID.

backendEnvironmentName :: Maybe Text

The name of the backend environment.

httpStatus :: Int

The response's http status code.

newDeleteBackendResponse

Arguments

:: Int
$sel:httpStatus:DeleteBackendResponse'

-> DeleteBackendResponse

Create a value of DeleteBackendResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DeleteBackendResponse', deleteBackendResponse_status - The current status of the request.

$sel:jobId:DeleteBackendResponse', deleteBackendResponse_jobId - The ID for the job.

$sel:operation:DeleteBackendResponse', deleteBackendResponse_operation - The name of the operation.

$sel:error:DeleteBackendResponse', deleteBackendResponse_error - If the request fails, this error is returned.

$sel:appId:DeleteBackend', deleteBackendResponse_appId - The app ID.

$sel:backendEnvironmentName:DeleteBackend', deleteBackendResponse_backendEnvironmentName - The name of the backend environment.

$sel:httpStatus:DeleteBackendResponse', deleteBackendResponse_httpStatus - The response's http status code.

Response Lenses

deleteBackendResponse_status :: Lens' DeleteBackendResponse (Maybe Text)

The current status of the request.

deleteBackendResponse_jobId :: Lens' DeleteBackendResponse (Maybe Text)

The ID for the job.

deleteBackendResponse_operation :: Lens' DeleteBackendResponse (Maybe Text)

The name of the operation.

deleteBackendResponse_error :: Lens' DeleteBackendResponse (Maybe Text)

If the request fails, this error is returned.

deleteBackendResponse_appId :: Lens' DeleteBackendResponse (Maybe Text)

The app ID.

deleteBackendResponse_backendEnvironmentName :: Lens' DeleteBackendResponse (Maybe Text)

The name of the backend environment.

deleteBackendResponse_httpStatus :: Lens' DeleteBackendResponse Int

The response's http status code.