Network.AWS.AmplifyBackend.DeleteBackendAuth

Description

Deletes an existing backend authentication resource.

Creating a Request

data DeleteBackendAuth

The request body for DeleteBackendAuth.

See: newDeleteBackendAuth smart constructor.

Constructors

DeleteBackendAuth'

Fields

appId :: Text

The app ID.

backendEnvironmentName :: Text

The name of the backend environment.

resourceName :: Text

The name of this resource.

newDeleteBackendAuth

Arguments

:: Text
$sel:appId:DeleteBackendAuth'

-> Text
$sel:backendEnvironmentName:DeleteBackendAuth'

-> Text
$sel:resourceName:DeleteBackendAuth'

-> DeleteBackendAuth

Create a value of DeleteBackendAuth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:DeleteBackendAuth', deleteBackendAuth_appId - The app ID.

$sel:backendEnvironmentName:DeleteBackendAuth', deleteBackendAuth_backendEnvironmentName - The name of the backend environment.

$sel:resourceName:DeleteBackendAuth', deleteBackendAuth_resourceName - The name of this resource.

Request Lenses

deleteBackendAuth_appId :: Lens' DeleteBackendAuth Text

The app ID.

deleteBackendAuth_backendEnvironmentName :: Lens' DeleteBackendAuth Text

The name of the backend environment.

deleteBackendAuth_resourceName :: Lens' DeleteBackendAuth Text

The name of this resource.

Destructuring the Response

data DeleteBackendAuthResponse

See: newDeleteBackendAuthResponse smart constructor.

Constructors

DeleteBackendAuthResponse'

Fields

status :: Maybe Text

The current status of the request.

jobId :: Maybe Text

The ID for the job.

operation :: Maybe Text

The name of the operation.

error :: Maybe Text

If the request fails, this error is returned.

appId :: Maybe Text

The app ID.

backendEnvironmentName :: Maybe Text

The name of the backend environment.

httpStatus :: Int

The response's http status code.

newDeleteBackendAuthResponse

Arguments

:: Int
$sel:httpStatus:DeleteBackendAuthResponse'

-> DeleteBackendAuthResponse

Create a value of DeleteBackendAuthResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DeleteBackendAuthResponse', deleteBackendAuthResponse_status - The current status of the request.

$sel:jobId:DeleteBackendAuthResponse', deleteBackendAuthResponse_jobId - The ID for the job.

$sel:operation:DeleteBackendAuthResponse', deleteBackendAuthResponse_operation - The name of the operation.

$sel:error:DeleteBackendAuthResponse', deleteBackendAuthResponse_error - If the request fails, this error is returned.

$sel:appId:DeleteBackendAuth', deleteBackendAuthResponse_appId - The app ID.

$sel:backendEnvironmentName:DeleteBackendAuth', deleteBackendAuthResponse_backendEnvironmentName - The name of the backend environment.

$sel:httpStatus:DeleteBackendAuthResponse', deleteBackendAuthResponse_httpStatus - The response's http status code.

Response Lenses

deleteBackendAuthResponse_status :: Lens' DeleteBackendAuthResponse (Maybe Text)

The current status of the request.

deleteBackendAuthResponse_jobId :: Lens' DeleteBackendAuthResponse (Maybe Text)

The ID for the job.

deleteBackendAuthResponse_operation :: Lens' DeleteBackendAuthResponse (Maybe Text)

The name of the operation.

deleteBackendAuthResponse_error :: Lens' DeleteBackendAuthResponse (Maybe Text)

If the request fails, this error is returned.

deleteBackendAuthResponse_appId :: Lens' DeleteBackendAuthResponse (Maybe Text)

The app ID.

deleteBackendAuthResponse_backendEnvironmentName :: Lens' DeleteBackendAuthResponse (Maybe Text)

The name of the backend environment.

deleteBackendAuthResponse_httpStatus :: Lens' DeleteBackendAuthResponse Int

The response's http status code.