Network.AWS.AmplifyBackend.GetBackend

Description

Provides project-level details for your Amplify UI project.

Creating a Request

data GetBackend

The request body for GetBackend.

See: newGetBackend smart constructor.

Constructors

GetBackend'

Fields

backendEnvironmentName :: Maybe Text

The name of the backend environment.

appId :: Text

The app ID.

newGetBackend

Arguments

:: Text
$sel:appId:GetBackend'

-> GetBackend

Create a value of GetBackend with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backendEnvironmentName:GetBackend', getBackend_backendEnvironmentName - The name of the backend environment.

$sel:appId:GetBackend', getBackend_appId - The app ID.

Request Lenses

getBackend_backendEnvironmentName :: Lens' GetBackend (Maybe Text)

The name of the backend environment.

getBackend_appId :: Lens' GetBackend Text

The app ID.

Destructuring the Response

data GetBackendResponse

See: newGetBackendResponse smart constructor.

Constructors

GetBackendResponse'

Fields

error :: Maybe Text

If the request failed, this is the returned error.

appName :: Maybe Text

The name of the app.

amplifyMetaConfig :: Maybe Text

A stringified version of the current configs for your Amplify project.

appId :: Maybe Text

The app ID.

amplifyFeatureFlags :: Maybe Text

A stringified version of the cli.json file for your Amplify project.

backendEnvironmentList :: Maybe [Text]

A list of backend environments in an array.

backendEnvironmentName :: Maybe Text

The name of the backend environment.

httpStatus :: Int

The response's http status code.

newGetBackendResponse

Arguments

:: Int
$sel:httpStatus:GetBackendResponse'

-> GetBackendResponse

Create a value of GetBackendResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:error:GetBackendResponse', getBackendResponse_error - If the request failed, this is the returned error.

$sel:appName:GetBackendResponse', getBackendResponse_appName - The name of the app.

$sel:amplifyMetaConfig:GetBackendResponse', getBackendResponse_amplifyMetaConfig - A stringified version of the current configs for your Amplify project.

$sel:appId:GetBackend', getBackendResponse_appId - The app ID.

$sel:amplifyFeatureFlags:GetBackendResponse', getBackendResponse_amplifyFeatureFlags - A stringified version of the cli.json file for your Amplify project.

$sel:backendEnvironmentList:GetBackendResponse', getBackendResponse_backendEnvironmentList - A list of backend environments in an array.

$sel:backendEnvironmentName:GetBackend', getBackendResponse_backendEnvironmentName - The name of the backend environment.

$sel:httpStatus:GetBackendResponse', getBackendResponse_httpStatus - The response's http status code.

Response Lenses

getBackendResponse_error :: Lens' GetBackendResponse (Maybe Text)

If the request failed, this is the returned error.

getBackendResponse_appName :: Lens' GetBackendResponse (Maybe Text)

The name of the app.

getBackendResponse_amplifyMetaConfig :: Lens' GetBackendResponse (Maybe Text)

A stringified version of the current configs for your Amplify project.

getBackendResponse_appId :: Lens' GetBackendResponse (Maybe Text)

The app ID.

getBackendResponse_amplifyFeatureFlags :: Lens' GetBackendResponse (Maybe Text)

A stringified version of the cli.json file for your Amplify project.

getBackendResponse_backendEnvironmentList :: Lens' GetBackendResponse (Maybe [Text])

A list of backend environments in an array.

getBackendResponse_backendEnvironmentName :: Lens' GetBackendResponse (Maybe Text)

The name of the backend environment.

getBackendResponse_httpStatus :: Lens' GetBackendResponse Int

The response's http status code.