Network.AWS.AmplifyBackend.GetBackendAuth

Description

Gets a backend auth details.

Creating a Request

data GetBackendAuth

The request body for GetBackendAuth.

See: newGetBackendAuth smart constructor.

Constructors

GetBackendAuth'

Fields

appId :: Text

The app ID.

backendEnvironmentName :: Text

The name of the backend environment.

resourceName :: Text

The name of this resource.

newGetBackendAuth

Arguments

:: Text
$sel:appId:GetBackendAuth'

-> Text
$sel:backendEnvironmentName:GetBackendAuth'

-> Text
$sel:resourceName:GetBackendAuth'

-> GetBackendAuth

Create a value of GetBackendAuth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:GetBackendAuth', getBackendAuth_appId - The app ID.

$sel:backendEnvironmentName:GetBackendAuth', getBackendAuth_backendEnvironmentName - The name of the backend environment.

$sel:resourceName:GetBackendAuth', getBackendAuth_resourceName - The name of this resource.

Request Lenses

getBackendAuth_appId :: Lens' GetBackendAuth Text

The app ID.

getBackendAuth_backendEnvironmentName :: Lens' GetBackendAuth Text

The name of the backend environment.

getBackendAuth_resourceName :: Lens' GetBackendAuth Text

The name of this resource.

Destructuring the Response

data GetBackendAuthResponse

See: newGetBackendAuthResponse smart constructor.

Constructors

GetBackendAuthResponse'

Fields

resourceName :: Maybe Text

The name of this resource.

error :: Maybe Text

If the request fails, this error is returned.

appId :: Maybe Text

The app ID.

resourceConfig :: Maybe CreateBackendAuthResourceConfig

The resource configuration for authorization requests to the backend of your Amplify project.

backendEnvironmentName :: Maybe Text

The name of the backend environment.

httpStatus :: Int

The response's http status code.

newGetBackendAuthResponse

Arguments

:: Int
$sel:httpStatus:GetBackendAuthResponse'

-> GetBackendAuthResponse

Create a value of GetBackendAuthResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceName:GetBackendAuth', getBackendAuthResponse_resourceName - The name of this resource.

$sel:error:GetBackendAuthResponse', getBackendAuthResponse_error - If the request fails, this error is returned.

$sel:appId:GetBackendAuth', getBackendAuthResponse_appId - The app ID.

$sel:resourceConfig:GetBackendAuthResponse', getBackendAuthResponse_resourceConfig - The resource configuration for authorization requests to the backend of your Amplify project.

$sel:backendEnvironmentName:GetBackendAuth', getBackendAuthResponse_backendEnvironmentName - The name of the backend environment.

$sel:httpStatus:GetBackendAuthResponse', getBackendAuthResponse_httpStatus - The response's http status code.

Response Lenses

getBackendAuthResponse_resourceName :: Lens' GetBackendAuthResponse (Maybe Text)

The name of this resource.

getBackendAuthResponse_error :: Lens' GetBackendAuthResponse (Maybe Text)

If the request fails, this error is returned.

getBackendAuthResponse_appId :: Lens' GetBackendAuthResponse (Maybe Text)

The app ID.

getBackendAuthResponse_resourceConfig :: Lens' GetBackendAuthResponse (Maybe CreateBackendAuthResourceConfig)

The resource configuration for authorization requests to the backend of your Amplify project.

getBackendAuthResponse_backendEnvironmentName :: Lens' GetBackendAuthResponse (Maybe Text)

The name of the backend environment.

getBackendAuthResponse_httpStatus :: Lens' GetBackendAuthResponse Int

The response's http status code.