Network.AWS.AmplifyBackend.ImportBackendAuth

Description

Imports an existing backend authentication resource.

Creating a Request

data ImportBackendAuth

The request body for ImportBackendAuth.

See: newImportBackendAuth smart constructor.

Constructors

ImportBackendAuth'

Fields

identityPoolId :: Maybe Text

The ID of the Amazon Cognito identity pool.

appId :: Text

The app ID.

backendEnvironmentName :: Text

The name of the backend environment.

userPoolId :: Text

The ID of the Amazon Cognito user pool.

nativeClientId :: Text

The ID of the Amazon Cognito native client.

webClientId :: Text

The ID of the Amazon Cognito web client.

newImportBackendAuth

Arguments

:: Text
$sel:appId:ImportBackendAuth'

-> Text
$sel:backendEnvironmentName:ImportBackendAuth'

-> Text
$sel:userPoolId:ImportBackendAuth'

-> Text
$sel:nativeClientId:ImportBackendAuth'

-> Text
$sel:webClientId:ImportBackendAuth'

-> ImportBackendAuth

Create a value of ImportBackendAuth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolId:ImportBackendAuth', importBackendAuth_identityPoolId - The ID of the Amazon Cognito identity pool.

$sel:appId:ImportBackendAuth', importBackendAuth_appId - The app ID.

$sel:backendEnvironmentName:ImportBackendAuth', importBackendAuth_backendEnvironmentName - The name of the backend environment.

$sel:userPoolId:ImportBackendAuth', importBackendAuth_userPoolId - The ID of the Amazon Cognito user pool.

$sel:nativeClientId:ImportBackendAuth', importBackendAuth_nativeClientId - The ID of the Amazon Cognito native client.

$sel:webClientId:ImportBackendAuth', importBackendAuth_webClientId - The ID of the Amazon Cognito web client.

Request Lenses

importBackendAuth_identityPoolId :: Lens' ImportBackendAuth (Maybe Text)

The ID of the Amazon Cognito identity pool.

importBackendAuth_appId :: Lens' ImportBackendAuth Text

The app ID.

importBackendAuth_backendEnvironmentName :: Lens' ImportBackendAuth Text

The name of the backend environment.

importBackendAuth_userPoolId :: Lens' ImportBackendAuth Text

The ID of the Amazon Cognito user pool.

importBackendAuth_nativeClientId :: Lens' ImportBackendAuth Text

The ID of the Amazon Cognito native client.

importBackendAuth_webClientId :: Lens' ImportBackendAuth Text

The ID of the Amazon Cognito web client.

Destructuring the Response

data ImportBackendAuthResponse

See: newImportBackendAuthResponse smart constructor.

Constructors

ImportBackendAuthResponse'

Fields

status :: Maybe Text

The current status of the request.

jobId :: Maybe Text

The ID for the job.

operation :: Maybe Text

The name of the operation.

error :: Maybe Text

If the request fails, this error is returned.

appId :: Maybe Text

The app ID.

backendEnvironmentName :: Maybe Text

The name of the backend environment.

httpStatus :: Int

The response's http status code.

newImportBackendAuthResponse

Arguments

:: Int
$sel:httpStatus:ImportBackendAuthResponse'

-> ImportBackendAuthResponse

Create a value of ImportBackendAuthResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ImportBackendAuthResponse', importBackendAuthResponse_status - The current status of the request.

$sel:jobId:ImportBackendAuthResponse', importBackendAuthResponse_jobId - The ID for the job.

$sel:operation:ImportBackendAuthResponse', importBackendAuthResponse_operation - The name of the operation.

$sel:error:ImportBackendAuthResponse', importBackendAuthResponse_error - If the request fails, this error is returned.

$sel:appId:ImportBackendAuth', importBackendAuthResponse_appId - The app ID.

$sel:backendEnvironmentName:ImportBackendAuth', importBackendAuthResponse_backendEnvironmentName - The name of the backend environment.

$sel:httpStatus:ImportBackendAuthResponse', importBackendAuthResponse_httpStatus - The response's http status code.

Response Lenses

importBackendAuthResponse_status :: Lens' ImportBackendAuthResponse (Maybe Text)

The current status of the request.

importBackendAuthResponse_jobId :: Lens' ImportBackendAuthResponse (Maybe Text)

The ID for the job.

importBackendAuthResponse_operation :: Lens' ImportBackendAuthResponse (Maybe Text)

The name of the operation.

importBackendAuthResponse_error :: Lens' ImportBackendAuthResponse (Maybe Text)

If the request fails, this error is returned.

importBackendAuthResponse_appId :: Lens' ImportBackendAuthResponse (Maybe Text)

The app ID.

importBackendAuthResponse_backendEnvironmentName :: Lens' ImportBackendAuthResponse (Maybe Text)

The name of the backend environment.

importBackendAuthResponse_httpStatus :: Lens' ImportBackendAuthResponse Int

The response's http status code.