Network.AWS.AmplifyBackend.Lens

Description

Operations

CloneBackend

cloneBackend_appId :: Lens' CloneBackend Text

The app ID.

cloneBackend_backendEnvironmentName :: Lens' CloneBackend Text

The name of the backend environment.

cloneBackend_targetEnvironmentName :: Lens' CloneBackend Text

The name of the destination backend environment to be created.

cloneBackendResponse_status :: Lens' CloneBackendResponse (Maybe Text)

The current status of the request.

cloneBackendResponse_jobId :: Lens' CloneBackendResponse (Maybe Text)

The ID for the job.

cloneBackendResponse_operation :: Lens' CloneBackendResponse (Maybe Text)

The name of the operation.

cloneBackendResponse_error :: Lens' CloneBackendResponse (Maybe Text)

If the request fails, this error is returned.

cloneBackendResponse_appId :: Lens' CloneBackendResponse (Maybe Text)

The app ID.

cloneBackendResponse_backendEnvironmentName :: Lens' CloneBackendResponse (Maybe Text)

The name of the backend environment.

cloneBackendResponse_httpStatus :: Lens' CloneBackendResponse Int

The response's http status code.

DeleteToken

deleteToken_sessionId :: Lens' DeleteToken Text

The session ID.

deleteToken_appId :: Lens' DeleteToken Text

The app ID.

deleteTokenResponse_isSuccess :: Lens' DeleteTokenResponse (Maybe Bool)

Indicates whether the request succeeded or failed.

deleteTokenResponse_httpStatus :: Lens' DeleteTokenResponse Int

The response's http status code.

CreateBackendConfig

createBackendConfig_backendManagerAppId :: Lens' CreateBackendConfig (Maybe Text)

The app ID for the backend manager.

createBackendConfig_appId :: Lens' CreateBackendConfig Text

The app ID.

createBackendConfigResponse_status :: Lens' CreateBackendConfigResponse (Maybe Text)

The current status of the request.

createBackendConfigResponse_jobId :: Lens' CreateBackendConfigResponse (Maybe Text)

The ID for the job.

createBackendConfigResponse_appId :: Lens' CreateBackendConfigResponse (Maybe Text)

The app ID.

createBackendConfigResponse_backendEnvironmentName :: Lens' CreateBackendConfigResponse (Maybe Text)

The name of the backend environment.

createBackendConfigResponse_httpStatus :: Lens' CreateBackendConfigResponse Int

The response's http status code.

ListBackendJobs

listBackendJobs_status :: Lens' ListBackendJobs (Maybe Text)

Filters the list of response objects to include only those with the specified status.

listBackendJobs_jobId :: Lens' ListBackendJobs (Maybe Text)

The ID for the job.

listBackendJobs_operation :: Lens' ListBackendJobs (Maybe Text)

Filters the list of response objects to include only those with the specified operation name.

listBackendJobs_nextToken :: Lens' ListBackendJobs (Maybe Text)

The token for the next set of results.

listBackendJobs_maxResults :: Lens' ListBackendJobs (Maybe Natural)

The maximum number of results that you want in the response.

listBackendJobs_appId :: Lens' ListBackendJobs Text

The app ID.

listBackendJobs_backendEnvironmentName :: Lens' ListBackendJobs Text

The name of the backend environment.

listBackendJobsResponse_nextToken :: Lens' ListBackendJobsResponse (Maybe Text)

The token for the next set of results.

listBackendJobsResponse_jobs :: Lens' ListBackendJobsResponse (Maybe [BackendJobRespObj])

An array of jobs and their properties.

listBackendJobsResponse_httpStatus :: Lens' ListBackendJobsResponse Int

The response's http status code.

GetBackend

getBackend_backendEnvironmentName :: Lens' GetBackend (Maybe Text)

The name of the backend environment.

getBackend_appId :: Lens' GetBackend Text

The app ID.

getBackendResponse_error :: Lens' GetBackendResponse (Maybe Text)

If the request failed, this is the returned error.

getBackendResponse_appName :: Lens' GetBackendResponse (Maybe Text)

The name of the app.

getBackendResponse_amplifyMetaConfig :: Lens' GetBackendResponse (Maybe Text)

A stringified version of the current configs for your Amplify project.

getBackendResponse_appId :: Lens' GetBackendResponse (Maybe Text)

The app ID.

getBackendResponse_amplifyFeatureFlags :: Lens' GetBackendResponse (Maybe Text)

A stringified version of the cli.json file for your Amplify project.

getBackendResponse_backendEnvironmentList :: Lens' GetBackendResponse (Maybe [Text])

A list of backend environments in an array.

getBackendResponse_backendEnvironmentName :: Lens' GetBackendResponse (Maybe Text)

The name of the backend environment.

getBackendResponse_httpStatus :: Lens' GetBackendResponse Int

The response's http status code.

GetBackendAuth

getBackendAuth_appId :: Lens' GetBackendAuth Text

The app ID.

getBackendAuth_backendEnvironmentName :: Lens' GetBackendAuth Text

The name of the backend environment.

getBackendAuth_resourceName :: Lens' GetBackendAuth Text

The name of this resource.

getBackendAuthResponse_resourceName :: Lens' GetBackendAuthResponse (Maybe Text)

The name of this resource.

getBackendAuthResponse_error :: Lens' GetBackendAuthResponse (Maybe Text)

If the request fails, this error is returned.

getBackendAuthResponse_appId :: Lens' GetBackendAuthResponse (Maybe Text)

The app ID.

getBackendAuthResponse_resourceConfig :: Lens' GetBackendAuthResponse (Maybe CreateBackendAuthResourceConfig)

The resource configuration for authorization requests to the backend of your Amplify project.

getBackendAuthResponse_backendEnvironmentName :: Lens' GetBackendAuthResponse (Maybe Text)

The name of the backend environment.

getBackendAuthResponse_httpStatus :: Lens' GetBackendAuthResponse Int

The response's http status code.

CreateBackendAuth

createBackendAuth_appId :: Lens' CreateBackendAuth Text

The app ID.

createBackendAuth_resourceName :: Lens' CreateBackendAuth Text

The name of this resource.

createBackendAuth_backendEnvironmentName :: Lens' CreateBackendAuth Text

The name of the backend environment.

createBackendAuth_resourceConfig :: Lens' CreateBackendAuth CreateBackendAuthResourceConfig

The resource configuration for this request object.

createBackendAuthResponse_status :: Lens' CreateBackendAuthResponse (Maybe Text)

The current status of the request.

createBackendAuthResponse_jobId :: Lens' CreateBackendAuthResponse (Maybe Text)

The ID for the job.

createBackendAuthResponse_operation :: Lens' CreateBackendAuthResponse (Maybe Text)

The name of the operation.

createBackendAuthResponse_error :: Lens' CreateBackendAuthResponse (Maybe Text)

If the request fails, this error is returned.

createBackendAuthResponse_appId :: Lens' CreateBackendAuthResponse (Maybe Text)

The app ID.

createBackendAuthResponse_backendEnvironmentName :: Lens' CreateBackendAuthResponse (Maybe Text)

The name of the backend environment.

createBackendAuthResponse_httpStatus :: Lens' CreateBackendAuthResponse Int

The response's http status code.

RemoveBackendConfig

removeBackendConfig_appId :: Lens' RemoveBackendConfig Text

The app ID.

removeBackendConfigResponse_error :: Lens' RemoveBackendConfigResponse (Maybe Text)

If the request fails, this error is returned.

removeBackendConfigResponse_httpStatus :: Lens' RemoveBackendConfigResponse Int

The response's http status code.

CreateBackend

createBackend_resourceName :: Lens' CreateBackend (Maybe Text)

The name of the resource.

createBackend_resourceConfig :: Lens' CreateBackend (Maybe ResourceConfig)

The resource configuration for the create backend request.

createBackend_appId :: Lens' CreateBackend Text

The app ID.

createBackend_backendEnvironmentName :: Lens' CreateBackend Text

The name of the backend environment.

createBackend_appName :: Lens' CreateBackend Text

The name of the app.

createBackendResponse_status :: Lens' CreateBackendResponse (Maybe Text)

The current status of the request.

createBackendResponse_jobId :: Lens' CreateBackendResponse (Maybe Text)

The ID for the job.

createBackendResponse_operation :: Lens' CreateBackendResponse (Maybe Text)

The name of the operation.

createBackendResponse_error :: Lens' CreateBackendResponse (Maybe Text)

If the request fails, this error is returned.

createBackendResponse_appId :: Lens' CreateBackendResponse (Maybe Text)

The app ID.

createBackendResponse_backendEnvironmentName :: Lens' CreateBackendResponse (Maybe Text)

The name of the backend environment.

createBackendResponse_httpStatus :: Lens' CreateBackendResponse Int

The response's http status code.

GenerateBackendAPIModels

generateBackendAPIModels_appId :: Lens' GenerateBackendAPIModels Text

The app ID.

generateBackendAPIModels_backendEnvironmentName :: Lens' GenerateBackendAPIModels Text

The name of the backend environment.

generateBackendAPIModels_resourceName :: Lens' GenerateBackendAPIModels Text

The name of this resource.

generateBackendAPIModelsResponse_status :: Lens' GenerateBackendAPIModelsResponse (Maybe Text)

The current status of the request.

generateBackendAPIModelsResponse_jobId :: Lens' GenerateBackendAPIModelsResponse (Maybe Text)

The ID for the job.

generateBackendAPIModelsResponse_operation :: Lens' GenerateBackendAPIModelsResponse (Maybe Text)

The name of the operation.

generateBackendAPIModelsResponse_error :: Lens' GenerateBackendAPIModelsResponse (Maybe Text)

If the request fails, this error is returned.

generateBackendAPIModelsResponse_appId :: Lens' GenerateBackendAPIModelsResponse (Maybe Text)

The app ID.

generateBackendAPIModelsResponse_backendEnvironmentName :: Lens' GenerateBackendAPIModelsResponse (Maybe Text)

The name of the backend environment.

generateBackendAPIModelsResponse_httpStatus :: Lens' GenerateBackendAPIModelsResponse Int

The response's http status code.

UpdateBackendConfig

updateBackendConfig_loginAuthConfig :: Lens' UpdateBackendConfig (Maybe LoginAuthConfigReqObj)

Describes the Amazon Cognito configuration for Admin UI access.

updateBackendConfig_appId :: Lens' UpdateBackendConfig Text

The app ID.

updateBackendConfigResponse_error :: Lens' UpdateBackendConfigResponse (Maybe Text)

If the request fails, this error is returned.

updateBackendConfigResponse_appId :: Lens' UpdateBackendConfigResponse (Maybe Text)

The app ID.

updateBackendConfigResponse_backendManagerAppId :: Lens' UpdateBackendConfigResponse (Maybe Text)

The app ID for the backend manager.

updateBackendConfigResponse_loginAuthConfig :: Lens' UpdateBackendConfigResponse (Maybe LoginAuthConfigReqObj)

Describes the Amazon Cognito configurations for the Admin UI auth resource to log in with.

updateBackendConfigResponse_httpStatus :: Lens' UpdateBackendConfigResponse Int

The response's http status code.

CreateToken

createToken_appId :: Lens' CreateToken Text

The app ID.

createTokenResponse_ttl :: Lens' CreateTokenResponse (Maybe Text)

The expiry time for the one-time generated token code.

createTokenResponse_appId :: Lens' CreateTokenResponse (Maybe Text)

The app ID.

createTokenResponse_sessionId :: Lens' CreateTokenResponse (Maybe Text)

A unique ID provided when creating a new challenge token.

createTokenResponse_challengeCode :: Lens' CreateTokenResponse (Maybe Text)

One-time challenge code for authenticating into the Amplify Admin UI.

createTokenResponse_httpStatus :: Lens' CreateTokenResponse Int

The response's http status code.

GetBackendAPI

getBackendAPI_resourceConfig :: Lens' GetBackendAPI (Maybe BackendAPIResourceConfig)

Defines the resource configuration for the data model in your Amplify project.

getBackendAPI_appId :: Lens' GetBackendAPI Text

The app ID.

getBackendAPI_backendEnvironmentName :: Lens' GetBackendAPI Text

The name of the backend environment.

getBackendAPI_resourceName :: Lens' GetBackendAPI Text

The name of this resource.

getBackendAPIResponse_resourceName :: Lens' GetBackendAPIResponse (Maybe Text)

The name of this resource.

getBackendAPIResponse_error :: Lens' GetBackendAPIResponse (Maybe Text)

If the request fails, this error is returned.

getBackendAPIResponse_appId :: Lens' GetBackendAPIResponse (Maybe Text)

The app ID.

getBackendAPIResponse_resourceConfig :: Lens' GetBackendAPIResponse (Maybe BackendAPIResourceConfig)

The resource configuration for this response object.

getBackendAPIResponse_backendEnvironmentName :: Lens' GetBackendAPIResponse (Maybe Text)

The name of the backend environment.

getBackendAPIResponse_httpStatus :: Lens' GetBackendAPIResponse Int

The response's http status code.

UpdateBackendJob

updateBackendJob_status :: Lens' UpdateBackendJob (Maybe Text)

Filters the list of response objects to include only those with the specified status.

updateBackendJob_operation :: Lens' UpdateBackendJob (Maybe Text)

Filters the list of response objects to include only those with the specified operation name.

updateBackendJob_appId :: Lens' UpdateBackendJob Text

The app ID.

updateBackendJob_backendEnvironmentName :: Lens' UpdateBackendJob Text

The name of the backend environment.

updateBackendJob_jobId :: Lens' UpdateBackendJob Text

The ID for the job.

updateBackendJobResponse_status :: Lens' UpdateBackendJobResponse (Maybe Text)

The current status of the request.

updateBackendJobResponse_jobId :: Lens' UpdateBackendJobResponse (Maybe Text)

The ID for the job.

updateBackendJobResponse_operation :: Lens' UpdateBackendJobResponse (Maybe Text)

The name of the operation.

updateBackendJobResponse_error :: Lens' UpdateBackendJobResponse (Maybe Text)

If the request fails, this error is returned.

updateBackendJobResponse_updateTime :: Lens' UpdateBackendJobResponse (Maybe Text)

The time when the job was last updated.

updateBackendJobResponse_appId :: Lens' UpdateBackendJobResponse (Maybe Text)

The app ID.

updateBackendJobResponse_backendEnvironmentName :: Lens' UpdateBackendJobResponse (Maybe Text)

The name of the backend environment.

updateBackendJobResponse_createTime :: Lens' UpdateBackendJobResponse (Maybe Text)

The time when the job was created.

updateBackendJobResponse_httpStatus :: Lens' UpdateBackendJobResponse Int

The response's http status code.

UpdateBackendAPI

updateBackendAPI_resourceConfig :: Lens' UpdateBackendAPI (Maybe BackendAPIResourceConfig)

Defines the resource configuration for the data model in your Amplify project.

updateBackendAPI_appId :: Lens' UpdateBackendAPI Text

The app ID.

updateBackendAPI_backendEnvironmentName :: Lens' UpdateBackendAPI Text

The name of the backend environment.

updateBackendAPI_resourceName :: Lens' UpdateBackendAPI Text

The name of this resource.

updateBackendAPIResponse_status :: Lens' UpdateBackendAPIResponse (Maybe Text)

The current status of the request.

updateBackendAPIResponse_jobId :: Lens' UpdateBackendAPIResponse (Maybe Text)

The ID for the job.

updateBackendAPIResponse_operation :: Lens' UpdateBackendAPIResponse (Maybe Text)

The name of the operation.

updateBackendAPIResponse_error :: Lens' UpdateBackendAPIResponse (Maybe Text)

If the request fails, this error is returned.

updateBackendAPIResponse_appId :: Lens' UpdateBackendAPIResponse (Maybe Text)

The app ID.

updateBackendAPIResponse_backendEnvironmentName :: Lens' UpdateBackendAPIResponse (Maybe Text)

The name of the backend environment.

updateBackendAPIResponse_httpStatus :: Lens' UpdateBackendAPIResponse Int

The response's http status code.

DeleteBackendAPI

deleteBackendAPI_resourceConfig :: Lens' DeleteBackendAPI (Maybe BackendAPIResourceConfig)

Defines the resource configuration for the data model in your Amplify project.

deleteBackendAPI_appId :: Lens' DeleteBackendAPI Text

The app ID.

deleteBackendAPI_backendEnvironmentName :: Lens' DeleteBackendAPI Text

The name of the backend environment.

deleteBackendAPI_resourceName :: Lens' DeleteBackendAPI Text

The name of this resource.

deleteBackendAPIResponse_status :: Lens' DeleteBackendAPIResponse (Maybe Text)

The current status of the request.

deleteBackendAPIResponse_jobId :: Lens' DeleteBackendAPIResponse (Maybe Text)

The ID for the job.

deleteBackendAPIResponse_operation :: Lens' DeleteBackendAPIResponse (Maybe Text)

The name of the operation.

deleteBackendAPIResponse_error :: Lens' DeleteBackendAPIResponse (Maybe Text)

If the request fails, this error is returned.

deleteBackendAPIResponse_appId :: Lens' DeleteBackendAPIResponse (Maybe Text)

The app ID.

deleteBackendAPIResponse_backendEnvironmentName :: Lens' DeleteBackendAPIResponse (Maybe Text)

The name of the backend environment.

deleteBackendAPIResponse_httpStatus :: Lens' DeleteBackendAPIResponse Int

The response's http status code.

GetBackendJob

getBackendJob_appId :: Lens' GetBackendJob Text

The app ID.

getBackendJob_backendEnvironmentName :: Lens' GetBackendJob Text

The name of the backend environment.

getBackendJob_jobId :: Lens' GetBackendJob Text

The ID for the job.

getBackendJobResponse_status :: Lens' GetBackendJobResponse (Maybe Text)

The current status of the request.

getBackendJobResponse_jobId :: Lens' GetBackendJobResponse (Maybe Text)

The ID for the job.

getBackendJobResponse_operation :: Lens' GetBackendJobResponse (Maybe Text)

The name of the operation.

getBackendJobResponse_error :: Lens' GetBackendJobResponse (Maybe Text)

If the request fails, this error is returned.

getBackendJobResponse_updateTime :: Lens' GetBackendJobResponse (Maybe Text)

The time when the job was last updated.

getBackendJobResponse_appId :: Lens' GetBackendJobResponse (Maybe Text)

The app ID.

getBackendJobResponse_backendEnvironmentName :: Lens' GetBackendJobResponse (Maybe Text)

The name of the backend environment.

getBackendJobResponse_createTime :: Lens' GetBackendJobResponse (Maybe Text)

The time when the job was created.

getBackendJobResponse_httpStatus :: Lens' GetBackendJobResponse Int

The response's http status code.

DeleteBackend

deleteBackend_appId :: Lens' DeleteBackend Text

The app ID.

deleteBackend_backendEnvironmentName :: Lens' DeleteBackend Text

The name of the backend environment.

deleteBackendResponse_status :: Lens' DeleteBackendResponse (Maybe Text)

The current status of the request.

deleteBackendResponse_jobId :: Lens' DeleteBackendResponse (Maybe Text)

The ID for the job.

deleteBackendResponse_operation :: Lens' DeleteBackendResponse (Maybe Text)

The name of the operation.

deleteBackendResponse_error :: Lens' DeleteBackendResponse (Maybe Text)

If the request fails, this error is returned.

deleteBackendResponse_appId :: Lens' DeleteBackendResponse (Maybe Text)

The app ID.

deleteBackendResponse_backendEnvironmentName :: Lens' DeleteBackendResponse (Maybe Text)

The name of the backend environment.

deleteBackendResponse_httpStatus :: Lens' DeleteBackendResponse Int

The response's http status code.

DeleteBackendAuth

deleteBackendAuth_appId :: Lens' DeleteBackendAuth Text

The app ID.

deleteBackendAuth_backendEnvironmentName :: Lens' DeleteBackendAuth Text

The name of the backend environment.

deleteBackendAuth_resourceName :: Lens' DeleteBackendAuth Text

The name of this resource.

deleteBackendAuthResponse_status :: Lens' DeleteBackendAuthResponse (Maybe Text)

The current status of the request.

deleteBackendAuthResponse_jobId :: Lens' DeleteBackendAuthResponse (Maybe Text)

The ID for the job.

deleteBackendAuthResponse_operation :: Lens' DeleteBackendAuthResponse (Maybe Text)

The name of the operation.

deleteBackendAuthResponse_error :: Lens' DeleteBackendAuthResponse (Maybe Text)

If the request fails, this error is returned.

deleteBackendAuthResponse_appId :: Lens' DeleteBackendAuthResponse (Maybe Text)

The app ID.

deleteBackendAuthResponse_backendEnvironmentName :: Lens' DeleteBackendAuthResponse (Maybe Text)

The name of the backend environment.

deleteBackendAuthResponse_httpStatus :: Lens' DeleteBackendAuthResponse Int

The response's http status code.

UpdateBackendAuth

updateBackendAuth_appId :: Lens' UpdateBackendAuth Text

The app ID.

updateBackendAuth_backendEnvironmentName :: Lens' UpdateBackendAuth Text

The name of the backend environment.

updateBackendAuth_resourceName :: Lens' UpdateBackendAuth Text

The name of this resource.

updateBackendAuth_resourceConfig :: Lens' UpdateBackendAuth UpdateBackendAuthResourceConfig

The resource configuration for this request object.

updateBackendAuthResponse_status :: Lens' UpdateBackendAuthResponse (Maybe Text)

The current status of the request.

updateBackendAuthResponse_jobId :: Lens' UpdateBackendAuthResponse (Maybe Text)

The ID for the job.

updateBackendAuthResponse_operation :: Lens' UpdateBackendAuthResponse (Maybe Text)

The name of the operation.

updateBackendAuthResponse_error :: Lens' UpdateBackendAuthResponse (Maybe Text)

If the request fails, this error is returned.

updateBackendAuthResponse_appId :: Lens' UpdateBackendAuthResponse (Maybe Text)

The app ID.

updateBackendAuthResponse_backendEnvironmentName :: Lens' UpdateBackendAuthResponse (Maybe Text)

The name of the backend environment.

updateBackendAuthResponse_httpStatus :: Lens' UpdateBackendAuthResponse Int

The response's http status code.

GetToken

getToken_sessionId :: Lens' GetToken Text

The session ID.

getToken_appId :: Lens' GetToken Text

The app ID.

getTokenResponse_ttl :: Lens' GetTokenResponse (Maybe Text)

The expiry time for the one-time generated token code.

getTokenResponse_appId :: Lens' GetTokenResponse (Maybe Text)

The app ID.

getTokenResponse_sessionId :: Lens' GetTokenResponse (Maybe Text)

A unique ID provided when creating a new challenge token.

getTokenResponse_challengeCode :: Lens' GetTokenResponse (Maybe Text)

The one-time challenge code for authenticating into the Amplify Admin UI.

getTokenResponse_httpStatus :: Lens' GetTokenResponse Int

The response's http status code.

RemoveAllBackends

removeAllBackends_cleanAmplifyApp :: Lens' RemoveAllBackends (Maybe Bool)

Cleans up the Amplify Console app if this value is set to true.

removeAllBackends_appId :: Lens' RemoveAllBackends Text

The app ID.

removeAllBackendsResponse_status :: Lens' RemoveAllBackendsResponse (Maybe Text)

The current status of the request.

removeAllBackendsResponse_jobId :: Lens' RemoveAllBackendsResponse (Maybe Text)

The ID for the job.

removeAllBackendsResponse_operation :: Lens' RemoveAllBackendsResponse (Maybe Text)

The name of the operation.

removeAllBackendsResponse_error :: Lens' RemoveAllBackendsResponse (Maybe Text)

If the request fails, this error is returned.

removeAllBackendsResponse_appId :: Lens' RemoveAllBackendsResponse (Maybe Text)

The app ID.

removeAllBackendsResponse_httpStatus :: Lens' RemoveAllBackendsResponse Int

The response's http status code.

CreateBackendAPI

createBackendAPI_appId :: Lens' CreateBackendAPI Text

The app ID.

createBackendAPI_resourceName :: Lens' CreateBackendAPI Text

The name of this resource.

createBackendAPI_backendEnvironmentName :: Lens' CreateBackendAPI Text

The name of the backend environment.

createBackendAPI_resourceConfig :: Lens' CreateBackendAPI BackendAPIResourceConfig

The resource configuration for this request.

createBackendAPIResponse_status :: Lens' CreateBackendAPIResponse (Maybe Text)

The current status of the request.

createBackendAPIResponse_jobId :: Lens' CreateBackendAPIResponse (Maybe Text)

The ID for the job.

createBackendAPIResponse_operation :: Lens' CreateBackendAPIResponse (Maybe Text)

The name of the operation.

createBackendAPIResponse_error :: Lens' CreateBackendAPIResponse (Maybe Text)

If the request fails, this error is returned.

createBackendAPIResponse_appId :: Lens' CreateBackendAPIResponse (Maybe Text)

The app ID.

createBackendAPIResponse_backendEnvironmentName :: Lens' CreateBackendAPIResponse (Maybe Text)

The name of the backend environment.

createBackendAPIResponse_httpStatus :: Lens' CreateBackendAPIResponse Int

The response's http status code.

GetBackendAPIModels

getBackendAPIModels_appId :: Lens' GetBackendAPIModels Text

The app ID.

getBackendAPIModels_backendEnvironmentName :: Lens' GetBackendAPIModels Text

The name of the backend environment.

getBackendAPIModels_resourceName :: Lens' GetBackendAPIModels Text

The name of this resource.

getBackendAPIModelsResponse_status :: Lens' GetBackendAPIModelsResponse (Maybe Status)

The current status of the request.

getBackendAPIModelsResponse_models :: Lens' GetBackendAPIModelsResponse (Maybe Text)

Stringified JSON of the datastore model.

getBackendAPIModelsResponse_httpStatus :: Lens' GetBackendAPIModelsResponse Int

The response's http status code.

ImportBackendAuth

importBackendAuth_identityPoolId :: Lens' ImportBackendAuth (Maybe Text)

The ID of the Amazon Cognito identity pool.

importBackendAuth_appId :: Lens' ImportBackendAuth Text

The app ID.

importBackendAuth_backendEnvironmentName :: Lens' ImportBackendAuth Text

The name of the backend environment.

importBackendAuth_userPoolId :: Lens' ImportBackendAuth Text

The ID of the Amazon Cognito user pool.

importBackendAuth_nativeClientId :: Lens' ImportBackendAuth Text

The ID of the Amazon Cognito native client.

importBackendAuth_webClientId :: Lens' ImportBackendAuth Text

The ID of the Amazon Cognito web client.

importBackendAuthResponse_status :: Lens' ImportBackendAuthResponse (Maybe Text)

The current status of the request.

importBackendAuthResponse_jobId :: Lens' ImportBackendAuthResponse (Maybe Text)

The ID for the job.

importBackendAuthResponse_operation :: Lens' ImportBackendAuthResponse (Maybe Text)

The name of the operation.

importBackendAuthResponse_error :: Lens' ImportBackendAuthResponse (Maybe Text)

If the request fails, this error is returned.

importBackendAuthResponse_appId :: Lens' ImportBackendAuthResponse (Maybe Text)

The app ID.

importBackendAuthResponse_backendEnvironmentName :: Lens' ImportBackendAuthResponse (Maybe Text)

The name of the backend environment.

importBackendAuthResponse_httpStatus :: Lens' ImportBackendAuthResponse Int

The response's http status code.

Types

BackendAPIAppSyncAuthSettings

backendAPIAppSyncAuthSettings_openIDProviderName :: Lens' BackendAPIAppSyncAuthSettings (Maybe Text)

The OpenID provider name, if OpenID was used as an authentication mechanism to access your data models.

backendAPIAppSyncAuthSettings_openIDClientId :: Lens' BackendAPIAppSyncAuthSettings (Maybe Text)

The clientID for openID, if openID was used as an authentication setting to access your data models.

backendAPIAppSyncAuthSettings_openIDIssueURL :: Lens' BackendAPIAppSyncAuthSettings (Maybe Text)

The openID issuer URL, if openID was used as an authentication setting to access your data models.

backendAPIAppSyncAuthSettings_cognitoUserPoolId :: Lens' BackendAPIAppSyncAuthSettings (Maybe Text)

The Amazon Cognito user pool ID, if Amazon Cognito was used as an authentication setting to access your data models.

backendAPIAppSyncAuthSettings_openIDAuthTTL :: Lens' BackendAPIAppSyncAuthSettings (Maybe Text)

The expiry time for the OpenID authentication mechanism.

backendAPIAppSyncAuthSettings_description :: Lens' BackendAPIAppSyncAuthSettings (Maybe Text)

The API key description for API_KEY, if it was used as an authentication mechanism to access your data models.

backendAPIAppSyncAuthSettings_expirationTime :: Lens' BackendAPIAppSyncAuthSettings (Maybe Double)

The API key expiration time for API_KEY, if it was used as an authentication mechanism to access your data models.

backendAPIAppSyncAuthSettings_openIDIatTTL :: Lens' BackendAPIAppSyncAuthSettings (Maybe Text)

The expiry time for the OpenID authentication mechanism.

BackendAPIAuthType

backendAPIAuthType_settings :: Lens' BackendAPIAuthType (Maybe BackendAPIAppSyncAuthSettings)

Describes settings for the authentication mode.

backendAPIAuthType_mode :: Lens' BackendAPIAuthType (Maybe Mode)

Describes the authentication mode.

BackendAPIConflictResolution

backendAPIConflictResolution_resolutionStrategy :: Lens' BackendAPIConflictResolution (Maybe ResolutionStrategy)

The strategy for conflict resolution.

BackendAPIResourceConfig

backendAPIResourceConfig_apiName :: Lens' BackendAPIResourceConfig (Maybe Text)

The API name used to interact with the data model, configured as a part of your Amplify project.

backendAPIResourceConfig_service :: Lens' BackendAPIResourceConfig (Maybe Text)

The service used to provision and interact with the data model.

backendAPIResourceConfig_transformSchema :: Lens' BackendAPIResourceConfig (Maybe Text)

The definition of the data model in the annotated transform of the GraphQL schema.

backendAPIResourceConfig_additionalAuthTypes :: Lens' BackendAPIResourceConfig (Maybe [BackendAPIAuthType])

Additional authentication methods used to interact with your data models.

backendAPIResourceConfig_conflictResolution :: Lens' BackendAPIResourceConfig (Maybe BackendAPIConflictResolution)

The conflict resolution strategy for your data stored in the data models.

backendAPIResourceConfig_defaultAuthType :: Lens' BackendAPIResourceConfig (Maybe BackendAPIAuthType)

The default authentication type for interacting with the configured data models in your Amplify project.

BackendAuthAppleProviderConfig

backendAuthAppleProviderConfig_clientId :: Lens' BackendAuthAppleProviderConfig (Maybe Text)

Describes the client_id (also called Services ID) that comes from Apple.

backendAuthAppleProviderConfig_privateKey :: Lens' BackendAuthAppleProviderConfig (Maybe Text)

Describes the private_key that comes from Apple.

backendAuthAppleProviderConfig_keyId :: Lens' BackendAuthAppleProviderConfig (Maybe Text)

Describes the key_id that comes from Apple.

backendAuthAppleProviderConfig_teamId :: Lens' BackendAuthAppleProviderConfig (Maybe Text)

Describes the team_id that comes from Apple.

BackendAuthSocialProviderConfig

backendAuthSocialProviderConfig_clientId :: Lens' BackendAuthSocialProviderConfig (Maybe Text)

Describes the client_id, which can be obtained from the third-party social federation provider.

backendAuthSocialProviderConfig_clientSecret :: Lens' BackendAuthSocialProviderConfig (Maybe Text)

Describes the client_secret, which can be obtained from third-party social federation providers.

BackendJobRespObj

backendJobRespObj_status :: Lens' BackendJobRespObj (Maybe Text)

The current status of the request.

backendJobRespObj_jobId :: Lens' BackendJobRespObj (Maybe Text)

The ID for the job.

backendJobRespObj_operation :: Lens' BackendJobRespObj (Maybe Text)

The name of the operation.

backendJobRespObj_error :: Lens' BackendJobRespObj (Maybe Text)

If the request fails, this error is returned.

backendJobRespObj_updateTime :: Lens' BackendJobRespObj (Maybe Text)

The time when the job was last updated.

backendJobRespObj_createTime :: Lens' BackendJobRespObj (Maybe Text)

The time when the job was created.

backendJobRespObj_appId :: Lens' BackendJobRespObj Text

The app ID.

backendJobRespObj_backendEnvironmentName :: Lens' BackendJobRespObj Text

The name of the backend environment.

CreateBackendAuthForgotPasswordConfig

createBackendAuthForgotPasswordConfig_emailSettings :: Lens' CreateBackendAuthForgotPasswordConfig (Maybe EmailSettings)

The configuration for the email sent when an app user forgets their password.

createBackendAuthForgotPasswordConfig_smsSettings :: Lens' CreateBackendAuthForgotPasswordConfig (Maybe SmsSettings)

The configuration for the SMS message sent when an app user forgets their password.

createBackendAuthForgotPasswordConfig_deliveryMethod :: Lens' CreateBackendAuthForgotPasswordConfig DeliveryMethod

Describes which mode to use (either SMS or email) to deliver messages to app users who want to recover their password.

CreateBackendAuthIdentityPoolConfig

createBackendAuthIdentityPoolConfig_unauthenticatedLogin :: Lens' CreateBackendAuthIdentityPoolConfig Bool

Set to true or false based on whether you want to enable guest authorization to your Amplify app.

createBackendAuthIdentityPoolConfig_identityPoolName :: Lens' CreateBackendAuthIdentityPoolConfig Text

Name of the Amazon Cognito identity pool used for authorization.

CreateBackendAuthMFAConfig

createBackendAuthMFAConfig_settings :: Lens' CreateBackendAuthMFAConfig (Maybe Settings)

Describes the configuration settings and methods for your Amplify app users to use MFA.

createBackendAuthMFAConfig_mfaMode :: Lens' CreateBackendAuthMFAConfig MFAMode

Describes whether MFA should be [ON, OFF, or OPTIONAL] for authentication in your Amplify project.

CreateBackendAuthOAuthConfig

createBackendAuthOAuthConfig_socialProviderSettings :: Lens' CreateBackendAuthOAuthConfig (Maybe SocialProviderSettings)

The settings for using social providers to access your Amplify app.

createBackendAuthOAuthConfig_domainPrefix :: Lens' CreateBackendAuthOAuthConfig (Maybe Text)

The domain prefix for your Amplify app.

createBackendAuthOAuthConfig_redirectSignOutURIs :: Lens' CreateBackendAuthOAuthConfig [Text]

Redirect URLs that OAuth uses when a user signs out of an Amplify app.

createBackendAuthOAuthConfig_redirectSignInURIs :: Lens' CreateBackendAuthOAuthConfig [Text]

The redirected URI for signing in to your Amplify app.

createBackendAuthOAuthConfig_oAuthGrantType :: Lens' CreateBackendAuthOAuthConfig OAuthGrantType

The OAuth grant type that you use to allow app users to authenticate from your Amplify app.

createBackendAuthOAuthConfig_oAuthScopes :: Lens' CreateBackendAuthOAuthConfig [OAuthScopesElement]

List of OAuth-related flows used to allow your app users to authenticate from your Amplify app.

CreateBackendAuthPasswordPolicyConfig

createBackendAuthPasswordPolicyConfig_additionalConstraints :: Lens' CreateBackendAuthPasswordPolicyConfig (Maybe [AdditionalConstraintsElement])

Additional constraints for the password used to access the backend of your Amplify project.

createBackendAuthPasswordPolicyConfig_minimumLength :: Lens' CreateBackendAuthPasswordPolicyConfig Double

The minimum length of the password used to access the backend of your Amplify project.

CreateBackendAuthResourceConfig

createBackendAuthResourceConfig_identityPoolConfigs :: Lens' CreateBackendAuthResourceConfig (Maybe CreateBackendAuthIdentityPoolConfig)

Describes the authorization configuration for the Amazon Cognito identity pool, provisioned as a part of your auth resource in the Amplify project.

createBackendAuthResourceConfig_authResources :: Lens' CreateBackendAuthResourceConfig AuthResources

Defines whether you want to configure only authentication or both authentication and authorization settings.

createBackendAuthResourceConfig_userPoolConfigs :: Lens' CreateBackendAuthResourceConfig CreateBackendAuthUserPoolConfig

Describes authentication configuration for the Amazon Cognito user pool, provisioned as a part of your auth resource in the Amplify project.

createBackendAuthResourceConfig_service :: Lens' CreateBackendAuthResourceConfig Service

Defines the service name to use when configuring an authentication resource in your Amplify project.

CreateBackendAuthUserPoolConfig

createBackendAuthUserPoolConfig_passwordPolicy :: Lens' CreateBackendAuthUserPoolConfig (Maybe CreateBackendAuthPasswordPolicyConfig)

Describes the password policy for your Amazon Cognito user pool, configured as a part of your Amplify project.

createBackendAuthUserPoolConfig_mfa :: Lens' CreateBackendAuthUserPoolConfig (Maybe CreateBackendAuthMFAConfig)

Describes whether to apply multi-factor authentication policies for your Amazon Cognito user pool configured as a part of your Amplify project.

createBackendAuthUserPoolConfig_forgotPassword :: Lens' CreateBackendAuthUserPoolConfig (Maybe CreateBackendAuthForgotPasswordConfig)

Describes the forgotten password policy for your Amazon Cognito user pool, configured as a part of your Amplify project.

createBackendAuthUserPoolConfig_oAuth :: Lens' CreateBackendAuthUserPoolConfig (Maybe CreateBackendAuthOAuthConfig)

Describes the OAuth policy and rules for your Amazon Cognito user pool, configured as a part of your Amplify project.

createBackendAuthUserPoolConfig_requiredSignUpAttributes :: Lens' CreateBackendAuthUserPoolConfig [RequiredSignUpAttributesElement]

The required attributes to sign up new users in the user pool.

createBackendAuthUserPoolConfig_signInMethod :: Lens' CreateBackendAuthUserPoolConfig SignInMethod

Describes the sign-in methods that your Amplify app users use to log in using the Amazon Cognito user pool, configured as a part of your Amplify project.

createBackendAuthUserPoolConfig_userPoolName :: Lens' CreateBackendAuthUserPoolConfig Text

The Amazon Cognito user pool name.

EmailSettings

emailSettings_emailSubject :: Lens' EmailSettings (Maybe Text)

The subject of the email.

emailSettings_emailMessage :: Lens' EmailSettings (Maybe Text)

The body of the email.

LoginAuthConfigReqObj

loginAuthConfigReqObj_awsUserPoolsWebClientId :: Lens' LoginAuthConfigReqObj (Maybe Text)

The web client ID for the Amazon Cognito user pools.

loginAuthConfigReqObj_awsCognitoIdentityPoolId :: Lens' LoginAuthConfigReqObj (Maybe Text)

The Amazon Cognito identity pool ID used for the Amplify Admin UI login authorization.

loginAuthConfigReqObj_awsUserPoolsId :: Lens' LoginAuthConfigReqObj (Maybe Text)

The Amazon Cognito user pool ID used for Amplify Admin UI login authentication.

loginAuthConfigReqObj_awsCognitoRegion :: Lens' LoginAuthConfigReqObj (Maybe Text)

The AWS Region for the Amplify Admin UI login.

ResourceConfig

Settings

settings_smsMessage :: Lens' Settings (Maybe Text)

The body of the SMS message.

settings_mfaTypes :: Lens' Settings (Maybe [MfaTypesElement])

The supported MFA types.

SmsSettings

smsSettings_smsMessage :: Lens' SmsSettings (Maybe Text)

The body of the SMS message.

SocialProviderSettings

socialProviderSettings_loginWithAmazon :: Lens' SocialProviderSettings (Maybe BackendAuthSocialProviderConfig)

Undocumented member.

socialProviderSettings_signInWithApple :: Lens' SocialProviderSettings (Maybe BackendAuthAppleProviderConfig)

Undocumented member.

socialProviderSettings_facebook :: Lens' SocialProviderSettings (Maybe BackendAuthSocialProviderConfig)

Undocumented member.

socialProviderSettings_google :: Lens' SocialProviderSettings (Maybe BackendAuthSocialProviderConfig)

Undocumented member.

UpdateBackendAuthForgotPasswordConfig

updateBackendAuthForgotPasswordConfig_emailSettings :: Lens' UpdateBackendAuthForgotPasswordConfig (Maybe EmailSettings)

The configuration for the email sent when an app user forgets their password.

updateBackendAuthForgotPasswordConfig_smsSettings :: Lens' UpdateBackendAuthForgotPasswordConfig (Maybe SmsSettings)

The configuration for the SMS message sent when an Amplify app user forgets their password.

updateBackendAuthForgotPasswordConfig_deliveryMethod :: Lens' UpdateBackendAuthForgotPasswordConfig (Maybe DeliveryMethod)

Describes which mode to use (either SMS or email) to deliver messages to app users that want to recover their password.

UpdateBackendAuthIdentityPoolConfig

updateBackendAuthIdentityPoolConfig_unauthenticatedLogin :: Lens' UpdateBackendAuthIdentityPoolConfig (Maybe Bool)

A boolean value that can be set to allow or disallow guest-level authorization into your Amplify app.

UpdateBackendAuthMFAConfig

updateBackendAuthMFAConfig_settings :: Lens' UpdateBackendAuthMFAConfig (Maybe Settings)

The settings of your MFA configuration for the backend of your Amplify project.

updateBackendAuthMFAConfig_mfaMode :: Lens' UpdateBackendAuthMFAConfig (Maybe MFAMode)

The MFA mode for the backend of your Amplify project.

UpdateBackendAuthOAuthConfig

updateBackendAuthOAuthConfig_socialProviderSettings :: Lens' UpdateBackendAuthOAuthConfig (Maybe SocialProviderSettings)

Describes third-party social federation configurations for allowing your users to sign in with OAuth.

updateBackendAuthOAuthConfig_domainPrefix :: Lens' UpdateBackendAuthOAuthConfig (Maybe Text)

The Amazon Cognito domain prefix used to create a hosted UI for authentication.

updateBackendAuthOAuthConfig_oAuthScopes :: Lens' UpdateBackendAuthOAuthConfig (Maybe [OAuthScopesElement])

The list of OAuth-related flows that can allow users to authenticate from your Amplify app.

updateBackendAuthOAuthConfig_oAuthGrantType :: Lens' UpdateBackendAuthOAuthConfig (Maybe OAuthGrantType)

The OAuth grant type to allow app users to authenticate from your Amplify app.

updateBackendAuthOAuthConfig_redirectSignOutURIs :: Lens' UpdateBackendAuthOAuthConfig (Maybe [Text])

Redirect URLs that OAuth uses when a user signs out of an Amplify app.

updateBackendAuthOAuthConfig_redirectSignInURIs :: Lens' UpdateBackendAuthOAuthConfig (Maybe [Text])

Redirect URLs that OAuth uses when a user signs in to an Amplify app.

UpdateBackendAuthPasswordPolicyConfig

updateBackendAuthPasswordPolicyConfig_additionalConstraints :: Lens' UpdateBackendAuthPasswordPolicyConfig (Maybe [AdditionalConstraintsElement])

Describes additional constraints on password requirements to sign in to the auth resource, configured as a part of your Amplify project.

updateBackendAuthPasswordPolicyConfig_minimumLength :: Lens' UpdateBackendAuthPasswordPolicyConfig (Maybe Double)

Describes the minimum length of the password required to sign in to the auth resource, configured as a part of your Amplify project.

UpdateBackendAuthResourceConfig

updateBackendAuthResourceConfig_identityPoolConfigs :: Lens' UpdateBackendAuthResourceConfig (Maybe UpdateBackendAuthIdentityPoolConfig)

Describes the authorization configuration for the Amazon Cognito identity pool, provisioned as a part of your auth resource in the Amplify project.

updateBackendAuthResourceConfig_authResources :: Lens' UpdateBackendAuthResourceConfig AuthResources

Defines the service name to use when configuring an authentication resource in your Amplify project.

updateBackendAuthResourceConfig_userPoolConfigs :: Lens' UpdateBackendAuthResourceConfig UpdateBackendAuthUserPoolConfig

Describes the authentication configuration for the Amazon Cognito user pool, provisioned as a part of your auth resource in the Amplify project.

updateBackendAuthResourceConfig_service :: Lens' UpdateBackendAuthResourceConfig Service

Defines the service name to use when configuring an authentication resource in your Amplify project.

UpdateBackendAuthUserPoolConfig

updateBackendAuthUserPoolConfig_passwordPolicy :: Lens' UpdateBackendAuthUserPoolConfig (Maybe UpdateBackendAuthPasswordPolicyConfig)

Describes the password policy for your Amazon Cognito user pool, configured as a part of your Amplify project.

updateBackendAuthUserPoolConfig_mfa :: Lens' UpdateBackendAuthUserPoolConfig (Maybe UpdateBackendAuthMFAConfig)

Describes whether to apply multi-factor authentication policies for your Amazon Cognito user pool configured as a part of your Amplify project.

updateBackendAuthUserPoolConfig_forgotPassword :: Lens' UpdateBackendAuthUserPoolConfig (Maybe UpdateBackendAuthForgotPasswordConfig)

Describes the forgot password policy for your Amazon Cognito user pool, configured as a part of your Amplify project.

updateBackendAuthUserPoolConfig_oAuth :: Lens' UpdateBackendAuthUserPoolConfig (Maybe UpdateBackendAuthOAuthConfig)

Describes the OAuth policy and rules for your Amazon Cognito user pool, configured as a part of your Amplify project.