Network.AWS.AmplifyBackend.RemoveAllBackends

Description

Removes all backend environments from your Amplify project.

Creating a Request

data RemoveAllBackends

The request body for RemoveAllBackends.

See: newRemoveAllBackends smart constructor.

Constructors

RemoveAllBackends'

Fields

cleanAmplifyApp :: Maybe Bool

Cleans up the Amplify Console app if this value is set to true.

appId :: Text

The app ID.

newRemoveAllBackends

Arguments

:: Text
$sel:appId:RemoveAllBackends'

-> RemoveAllBackends

Create a value of RemoveAllBackends with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cleanAmplifyApp:RemoveAllBackends', removeAllBackends_cleanAmplifyApp - Cleans up the Amplify Console app if this value is set to true.

$sel:appId:RemoveAllBackends', removeAllBackends_appId - The app ID.

Request Lenses

removeAllBackends_cleanAmplifyApp :: Lens' RemoveAllBackends (Maybe Bool)

Cleans up the Amplify Console app if this value is set to true.

removeAllBackends_appId :: Lens' RemoveAllBackends Text

The app ID.

Destructuring the Response

data RemoveAllBackendsResponse

See: newRemoveAllBackendsResponse smart constructor.

Constructors

RemoveAllBackendsResponse'

Fields

status :: Maybe Text

The current status of the request.

jobId :: Maybe Text

The ID for the job.

operation :: Maybe Text

The name of the operation.

error :: Maybe Text

If the request fails, this error is returned.

appId :: Maybe Text

The app ID.

httpStatus :: Int

The response's http status code.

newRemoveAllBackendsResponse

Arguments

:: Int
$sel:httpStatus:RemoveAllBackendsResponse'

-> RemoveAllBackendsResponse

Create a value of RemoveAllBackendsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RemoveAllBackendsResponse', removeAllBackendsResponse_status - The current status of the request.

$sel:jobId:RemoveAllBackendsResponse', removeAllBackendsResponse_jobId - The ID for the job.

$sel:operation:RemoveAllBackendsResponse', removeAllBackendsResponse_operation - The name of the operation.

$sel:error:RemoveAllBackendsResponse', removeAllBackendsResponse_error - If the request fails, this error is returned.

$sel:appId:RemoveAllBackends', removeAllBackendsResponse_appId - The app ID.

$sel:httpStatus:RemoveAllBackendsResponse', removeAllBackendsResponse_httpStatus - The response's http status code.

Response Lenses

removeAllBackendsResponse_status :: Lens' RemoveAllBackendsResponse (Maybe Text)

The current status of the request.

removeAllBackendsResponse_jobId :: Lens' RemoveAllBackendsResponse (Maybe Text)

The ID for the job.

removeAllBackendsResponse_operation :: Lens' RemoveAllBackendsResponse (Maybe Text)

The name of the operation.

removeAllBackendsResponse_error :: Lens' RemoveAllBackendsResponse (Maybe Text)

If the request fails, this error is returned.

removeAllBackendsResponse_appId :: Lens' RemoveAllBackendsResponse (Maybe Text)

The app ID.

removeAllBackendsResponse_httpStatus :: Lens' RemoveAllBackendsResponse Int

The response's http status code.