Network.AWS.AmplifyBackend.Types.CreateBackendAuthResourceConfig

Description

Documentation

data CreateBackendAuthResourceConfig

Defines the resource configuration when creating an auth resource in your Amplify project.

See: newCreateBackendAuthResourceConfig smart constructor.

Constructors

CreateBackendAuthResourceConfig'

Fields

identityPoolConfigs :: Maybe CreateBackendAuthIdentityPoolConfig

Describes the authorization configuration for the Amazon Cognito identity pool, provisioned as a part of your auth resource in the Amplify project.

authResources :: AuthResources

Defines whether you want to configure only authentication or both authentication and authorization settings.

userPoolConfigs :: CreateBackendAuthUserPoolConfig

Describes authentication configuration for the Amazon Cognito user pool, provisioned as a part of your auth resource in the Amplify project.

service :: Service

Defines the service name to use when configuring an authentication resource in your Amplify project.

newCreateBackendAuthResourceConfig

Arguments

:: AuthResources
$sel:authResources:CreateBackendAuthResourceConfig'

-> CreateBackendAuthUserPoolConfig
$sel:userPoolConfigs:CreateBackendAuthResourceConfig'

-> Service
$sel:service:CreateBackendAuthResourceConfig'

-> CreateBackendAuthResourceConfig

Create a value of CreateBackendAuthResourceConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolConfigs:CreateBackendAuthResourceConfig', createBackendAuthResourceConfig_identityPoolConfigs - Describes the authorization configuration for the Amazon Cognito identity pool, provisioned as a part of your auth resource in the Amplify project.

$sel:authResources:CreateBackendAuthResourceConfig', createBackendAuthResourceConfig_authResources - Defines whether you want to configure only authentication or both authentication and authorization settings.

$sel:userPoolConfigs:CreateBackendAuthResourceConfig', createBackendAuthResourceConfig_userPoolConfigs - Describes authentication configuration for the Amazon Cognito user pool, provisioned as a part of your auth resource in the Amplify project.

$sel:service:CreateBackendAuthResourceConfig', createBackendAuthResourceConfig_service - Defines the service name to use when configuring an authentication resource in your Amplify project.

createBackendAuthResourceConfig_identityPoolConfigs :: Lens' CreateBackendAuthResourceConfig (Maybe CreateBackendAuthIdentityPoolConfig)

Describes the authorization configuration for the Amazon Cognito identity pool, provisioned as a part of your auth resource in the Amplify project.

createBackendAuthResourceConfig_authResources :: Lens' CreateBackendAuthResourceConfig AuthResources

Defines whether you want to configure only authentication or both authentication and authorization settings.

createBackendAuthResourceConfig_userPoolConfigs :: Lens' CreateBackendAuthResourceConfig CreateBackendAuthUserPoolConfig

Describes authentication configuration for the Amazon Cognito user pool, provisioned as a part of your auth resource in the Amplify project.

createBackendAuthResourceConfig_service :: Lens' CreateBackendAuthResourceConfig Service

Defines the service name to use when configuring an authentication resource in your Amplify project.