Network.AWS.AmplifyBackend.Types.Settings

Description

Documentation

data Settings

The settings of your MFA configuration for the backend of your Amplify project.

See: newSettings smart constructor.

Constructors

Settings'

Fields

smsMessage :: Maybe Text

The body of the SMS message.

mfaTypes :: Maybe [MfaTypesElement]

The supported MFA types.

newSettings :: Settings

Create a value of Settings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:smsMessage:Settings', settings_smsMessage - The body of the SMS message.

$sel:mfaTypes:Settings', settings_mfaTypes - The supported MFA types.

settings_smsMessage :: Lens' Settings (Maybe Text)

The body of the SMS message.

settings_mfaTypes :: Lens' Settings (Maybe [MfaTypesElement])

The supported MFA types.