Network.AWS.AmplifyBackend.Types.UpdateBackendAuthOAuthConfig

Description

Documentation

data UpdateBackendAuthOAuthConfig

The OAuth configurations for authenticating users into your Amplify app.

See: newUpdateBackendAuthOAuthConfig smart constructor.

Constructors

UpdateBackendAuthOAuthConfig'

Fields

socialProviderSettings :: Maybe SocialProviderSettings

Describes third-party social federation configurations for allowing your users to sign in with OAuth.

domainPrefix :: Maybe Text

The Amazon Cognito domain prefix used to create a hosted UI for authentication.

oAuthScopes :: Maybe [OAuthScopesElement]

The list of OAuth-related flows that can allow users to authenticate from your Amplify app.

oAuthGrantType :: Maybe OAuthGrantType

The OAuth grant type to allow app users to authenticate from your Amplify app.

redirectSignOutURIs :: Maybe [Text]

Redirect URLs that OAuth uses when a user signs out of an Amplify app.

redirectSignInURIs :: Maybe [Text]

Redirect URLs that OAuth uses when a user signs in to an Amplify app.

newUpdateBackendAuthOAuthConfig :: UpdateBackendAuthOAuthConfig

Create a value of UpdateBackendAuthOAuthConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:socialProviderSettings:UpdateBackendAuthOAuthConfig', updateBackendAuthOAuthConfig_socialProviderSettings - Describes third-party social federation configurations for allowing your users to sign in with OAuth.

$sel:domainPrefix:UpdateBackendAuthOAuthConfig', updateBackendAuthOAuthConfig_domainPrefix - The Amazon Cognito domain prefix used to create a hosted UI for authentication.

$sel:oAuthScopes:UpdateBackendAuthOAuthConfig', updateBackendAuthOAuthConfig_oAuthScopes - The list of OAuth-related flows that can allow users to authenticate from your Amplify app.

$sel:oAuthGrantType:UpdateBackendAuthOAuthConfig', updateBackendAuthOAuthConfig_oAuthGrantType - The OAuth grant type to allow app users to authenticate from your Amplify app.

$sel:redirectSignOutURIs:UpdateBackendAuthOAuthConfig', updateBackendAuthOAuthConfig_redirectSignOutURIs - Redirect URLs that OAuth uses when a user signs out of an Amplify app.

$sel:redirectSignInURIs:UpdateBackendAuthOAuthConfig', updateBackendAuthOAuthConfig_redirectSignInURIs - Redirect URLs that OAuth uses when a user signs in to an Amplify app.

updateBackendAuthOAuthConfig_socialProviderSettings :: Lens' UpdateBackendAuthOAuthConfig (Maybe SocialProviderSettings)

Describes third-party social federation configurations for allowing your users to sign in with OAuth.

updateBackendAuthOAuthConfig_domainPrefix :: Lens' UpdateBackendAuthOAuthConfig (Maybe Text)

The Amazon Cognito domain prefix used to create a hosted UI for authentication.

updateBackendAuthOAuthConfig_oAuthScopes :: Lens' UpdateBackendAuthOAuthConfig (Maybe [OAuthScopesElement])

The list of OAuth-related flows that can allow users to authenticate from your Amplify app.

updateBackendAuthOAuthConfig_oAuthGrantType :: Lens' UpdateBackendAuthOAuthConfig (Maybe OAuthGrantType)

The OAuth grant type to allow app users to authenticate from your Amplify app.

updateBackendAuthOAuthConfig_redirectSignOutURIs :: Lens' UpdateBackendAuthOAuthConfig (Maybe [Text])

Redirect URLs that OAuth uses when a user signs out of an Amplify app.

updateBackendAuthOAuthConfig_redirectSignInURIs :: Lens' UpdateBackendAuthOAuthConfig (Maybe [Text])

Redirect URLs that OAuth uses when a user signs in to an Amplify app.