Network.AWS.AmplifyBackend.Types.UpdateBackendAuthResourceConfig

Description

Documentation

data UpdateBackendAuthResourceConfig

Defines the resource configuration when updating an authentication resource in your Amplify project.

See: newUpdateBackendAuthResourceConfig smart constructor.

Constructors

UpdateBackendAuthResourceConfig'

Fields

identityPoolConfigs :: Maybe UpdateBackendAuthIdentityPoolConfig

Describes the authorization configuration for the Amazon Cognito identity pool, provisioned as a part of your auth resource in the Amplify project.

authResources :: AuthResources

Defines the service name to use when configuring an authentication resource in your Amplify project.

userPoolConfigs :: UpdateBackendAuthUserPoolConfig

Describes the authentication configuration for the Amazon Cognito user pool, provisioned as a part of your auth resource in the Amplify project.

service :: Service

Defines the service name to use when configuring an authentication resource in your Amplify project.

newUpdateBackendAuthResourceConfig

Arguments

:: AuthResources
$sel:authResources:UpdateBackendAuthResourceConfig'

-> UpdateBackendAuthUserPoolConfig
$sel:userPoolConfigs:UpdateBackendAuthResourceConfig'

-> Service
$sel:service:UpdateBackendAuthResourceConfig'

-> UpdateBackendAuthResourceConfig

Create a value of UpdateBackendAuthResourceConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identityPoolConfigs:UpdateBackendAuthResourceConfig', updateBackendAuthResourceConfig_identityPoolConfigs - Describes the authorization configuration for the Amazon Cognito identity pool, provisioned as a part of your auth resource in the Amplify project.

$sel:authResources:UpdateBackendAuthResourceConfig', updateBackendAuthResourceConfig_authResources - Defines the service name to use when configuring an authentication resource in your Amplify project.

$sel:userPoolConfigs:UpdateBackendAuthResourceConfig', updateBackendAuthResourceConfig_userPoolConfigs - Describes the authentication configuration for the Amazon Cognito user pool, provisioned as a part of your auth resource in the Amplify project.

$sel:service:UpdateBackendAuthResourceConfig', updateBackendAuthResourceConfig_service - Defines the service name to use when configuring an authentication resource in your Amplify project.

updateBackendAuthResourceConfig_identityPoolConfigs :: Lens' UpdateBackendAuthResourceConfig (Maybe UpdateBackendAuthIdentityPoolConfig)

Describes the authorization configuration for the Amazon Cognito identity pool, provisioned as a part of your auth resource in the Amplify project.

updateBackendAuthResourceConfig_authResources :: Lens' UpdateBackendAuthResourceConfig AuthResources

Defines the service name to use when configuring an authentication resource in your Amplify project.

updateBackendAuthResourceConfig_userPoolConfigs :: Lens' UpdateBackendAuthResourceConfig UpdateBackendAuthUserPoolConfig

Describes the authentication configuration for the Amazon Cognito user pool, provisioned as a part of your auth resource in the Amplify project.

updateBackendAuthResourceConfig_service :: Lens' UpdateBackendAuthResourceConfig Service

Defines the service name to use when configuring an authentication resource in your Amplify project.