Network.AWS.AmplifyBackend.UpdateBackendAPI

Description

Updates an existing backend API resource.

Creating a Request

data UpdateBackendAPI

The request body for UpdateBackendAPI.

See: newUpdateBackendAPI smart constructor.

Constructors

UpdateBackendAPI'

Fields

resourceConfig :: Maybe BackendAPIResourceConfig

Defines the resource configuration for the data model in your Amplify project.

appId :: Text

The app ID.

backendEnvironmentName :: Text

The name of the backend environment.

resourceName :: Text

The name of this resource.

newUpdateBackendAPI

Arguments

:: Text
$sel:appId:UpdateBackendAPI'

-> Text
$sel:backendEnvironmentName:UpdateBackendAPI'

-> Text
$sel:resourceName:UpdateBackendAPI'

-> UpdateBackendAPI

Create a value of UpdateBackendAPI with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceConfig:UpdateBackendAPI', updateBackendAPI_resourceConfig - Defines the resource configuration for the data model in your Amplify project.

$sel:appId:UpdateBackendAPI', updateBackendAPI_appId - The app ID.

$sel:backendEnvironmentName:UpdateBackendAPI', updateBackendAPI_backendEnvironmentName - The name of the backend environment.

$sel:resourceName:UpdateBackendAPI', updateBackendAPI_resourceName - The name of this resource.

Request Lenses

updateBackendAPI_resourceConfig :: Lens' UpdateBackendAPI (Maybe BackendAPIResourceConfig)

Defines the resource configuration for the data model in your Amplify project.

updateBackendAPI_appId :: Lens' UpdateBackendAPI Text

The app ID.

updateBackendAPI_backendEnvironmentName :: Lens' UpdateBackendAPI Text

The name of the backend environment.

updateBackendAPI_resourceName :: Lens' UpdateBackendAPI Text

The name of this resource.

Destructuring the Response

data UpdateBackendAPIResponse

See: newUpdateBackendAPIResponse smart constructor.

Constructors

UpdateBackendAPIResponse'

Fields

status :: Maybe Text

The current status of the request.

jobId :: Maybe Text

The ID for the job.

operation :: Maybe Text

The name of the operation.

error :: Maybe Text

If the request fails, this error is returned.

appId :: Maybe Text

The app ID.

backendEnvironmentName :: Maybe Text

The name of the backend environment.

httpStatus :: Int

The response's http status code.

newUpdateBackendAPIResponse

Arguments

:: Int
$sel:httpStatus:UpdateBackendAPIResponse'

-> UpdateBackendAPIResponse

Create a value of UpdateBackendAPIResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateBackendAPIResponse', updateBackendAPIResponse_status - The current status of the request.

$sel:jobId:UpdateBackendAPIResponse', updateBackendAPIResponse_jobId - The ID for the job.

$sel:operation:UpdateBackendAPIResponse', updateBackendAPIResponse_operation - The name of the operation.

$sel:error:UpdateBackendAPIResponse', updateBackendAPIResponse_error - If the request fails, this error is returned.

$sel:appId:UpdateBackendAPI', updateBackendAPIResponse_appId - The app ID.

$sel:backendEnvironmentName:UpdateBackendAPI', updateBackendAPIResponse_backendEnvironmentName - The name of the backend environment.

$sel:httpStatus:UpdateBackendAPIResponse', updateBackendAPIResponse_httpStatus - The response's http status code.

Response Lenses

updateBackendAPIResponse_status :: Lens' UpdateBackendAPIResponse (Maybe Text)

The current status of the request.

updateBackendAPIResponse_jobId :: Lens' UpdateBackendAPIResponse (Maybe Text)

The ID for the job.

updateBackendAPIResponse_operation :: Lens' UpdateBackendAPIResponse (Maybe Text)

The name of the operation.

updateBackendAPIResponse_error :: Lens' UpdateBackendAPIResponse (Maybe Text)

If the request fails, this error is returned.

updateBackendAPIResponse_appId :: Lens' UpdateBackendAPIResponse (Maybe Text)

The app ID.

updateBackendAPIResponse_backendEnvironmentName :: Lens' UpdateBackendAPIResponse (Maybe Text)

The name of the backend environment.

updateBackendAPIResponse_httpStatus :: Lens' UpdateBackendAPIResponse Int

The response's http status code.