Network.AWS.ApiGatewayV2.CreateRouteResponse

Description

Creates a RouteResponse for a Route.

Creating a Request

data CreateRouteResponse

Creates a new RouteResponse resource to represent a route response.

See: newCreateRouteResponse smart constructor.

Constructors

CreateRouteResponse'

Fields

modelSelectionExpression :: Maybe Text

The model selection expression for the route response. Supported only for WebSocket APIs.

responseModels :: Maybe (HashMap Text Text)

The response models for the route response.

responseParameters :: Maybe (HashMap Text ParameterConstraints)

The route response parameters.

apiId :: Text

The API identifier.

routeId :: Text

The route ID.

routeResponseKey :: Text

The route response key.

newCreateRouteResponse

Arguments

:: Text
$sel:apiId:CreateRouteResponse'

-> Text
$sel:routeId:CreateRouteResponse'

-> Text
$sel:routeResponseKey:CreateRouteResponse'

-> CreateRouteResponse

Create a value of CreateRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelSelectionExpression:CreateRouteResponse', createRouteResponse_modelSelectionExpression - The model selection expression for the route response. Supported only for WebSocket APIs.

$sel:responseModels:CreateRouteResponse', createRouteResponse_responseModels - The response models for the route response.

$sel:responseParameters:CreateRouteResponse', createRouteResponse_responseParameters - The route response parameters.

$sel:apiId:CreateRouteResponse', createRouteResponse_apiId - The API identifier.

$sel:routeId:CreateRouteResponse', createRouteResponse_routeId - The route ID.

$sel:routeResponseKey:CreateRouteResponse', createRouteResponse_routeResponseKey - The route response key.

Request Lenses

createRouteResponse_modelSelectionExpression :: Lens' CreateRouteResponse (Maybe Text)

The model selection expression for the route response. Supported only for WebSocket APIs.

createRouteResponse_responseModels :: Lens' CreateRouteResponse (Maybe (HashMap Text Text))

The response models for the route response.

createRouteResponse_responseParameters :: Lens' CreateRouteResponse (Maybe (HashMap Text ParameterConstraints))

The route response parameters.

createRouteResponse_apiId :: Lens' CreateRouteResponse Text

The API identifier.

createRouteResponse_routeId :: Lens' CreateRouteResponse Text

The route ID.

createRouteResponse_routeResponseKey :: Lens' CreateRouteResponse Text

The route response key.

Destructuring the Response

data CreateRouteResponseResponse

See: newCreateRouteResponseResponse smart constructor.

Constructors

CreateRouteResponseResponse'

Fields

modelSelectionExpression :: Maybe Text

Represents the model selection expression of a route response. Supported only for WebSocket APIs.

responseModels :: Maybe (HashMap Text Text)

Represents the response models of a route response.

routeResponseId :: Maybe Text

Represents the identifier of a route response.

routeResponseKey :: Maybe Text

Represents the route response key of a route response.

responseParameters :: Maybe (HashMap Text ParameterConstraints)

Represents the response parameters of a route response.

httpStatus :: Int

The response's http status code.

newCreateRouteResponseResponse

Arguments

:: Int
$sel:httpStatus:CreateRouteResponseResponse'

-> CreateRouteResponseResponse

Create a value of CreateRouteResponseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelSelectionExpression:CreateRouteResponse', createRouteResponseResponse_modelSelectionExpression - Represents the model selection expression of a route response. Supported only for WebSocket APIs.

$sel:responseModels:CreateRouteResponse', createRouteResponseResponse_responseModels - Represents the response models of a route response.

$sel:routeResponseId:CreateRouteResponseResponse', createRouteResponseResponse_routeResponseId - Represents the identifier of a route response.

$sel:routeResponseKey:CreateRouteResponse', createRouteResponseResponse_routeResponseKey - Represents the route response key of a route response.

$sel:responseParameters:CreateRouteResponse', createRouteResponseResponse_responseParameters - Represents the response parameters of a route response.

$sel:httpStatus:CreateRouteResponseResponse', createRouteResponseResponse_httpStatus - The response's http status code.

Response Lenses

createRouteResponseResponse_modelSelectionExpression :: Lens' CreateRouteResponseResponse (Maybe Text)

Represents the model selection expression of a route response. Supported only for WebSocket APIs.

createRouteResponseResponse_responseModels :: Lens' CreateRouteResponseResponse (Maybe (HashMap Text Text))

Represents the response models of a route response.

createRouteResponseResponse_routeResponseId :: Lens' CreateRouteResponseResponse (Maybe Text)

Represents the identifier of a route response.

createRouteResponseResponse_routeResponseKey :: Lens' CreateRouteResponseResponse (Maybe Text)

Represents the route response key of a route response.

createRouteResponseResponse_responseParameters :: Lens' CreateRouteResponseResponse (Maybe (HashMap Text ParameterConstraints))

Represents the response parameters of a route response.

createRouteResponseResponse_httpStatus :: Lens' CreateRouteResponseResponse Int

The response's http status code.