Network.AWS.ApiGatewayV2.DeleteAccessLogSettings

Description

Deletes the AccessLogSettings for a Stage. To disable access logging for a Stage, delete its AccessLogSettings.

Creating a Request

data DeleteAccessLogSettings

See: newDeleteAccessLogSettings smart constructor.

Constructors

DeleteAccessLogSettings'

Fields

stageName :: Text

The stage name. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

apiId :: Text

The API identifier.

newDeleteAccessLogSettings

Arguments

:: Text
$sel:stageName:DeleteAccessLogSettings'

-> Text
$sel:apiId:DeleteAccessLogSettings'

-> DeleteAccessLogSettings

Create a value of DeleteAccessLogSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stageName:DeleteAccessLogSettings', deleteAccessLogSettings_stageName - The stage name. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

$sel:apiId:DeleteAccessLogSettings', deleteAccessLogSettings_apiId - The API identifier.

Request Lenses

deleteAccessLogSettings_stageName :: Lens' DeleteAccessLogSettings Text

The stage name. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

deleteAccessLogSettings_apiId :: Lens' DeleteAccessLogSettings Text

The API identifier.

Destructuring the Response

data DeleteAccessLogSettingsResponse

See: newDeleteAccessLogSettingsResponse smart constructor.

Constructors

DeleteAccessLogSettingsResponse'

newDeleteAccessLogSettingsResponse :: DeleteAccessLogSettingsResponse

Create a value of DeleteAccessLogSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.