Network.AWS.ApiGatewayV2.GetApiMapping

Description

Gets an API mapping.

Creating a Request

data GetApiMapping

See: newGetApiMapping smart constructor.

Constructors

GetApiMapping'

Fields

apiMappingId :: Text

The API mapping identifier.

domainName :: Text

The domain name.

newGetApiMapping

Arguments

:: Text
$sel:apiMappingId:GetApiMapping'

-> Text
$sel:domainName:GetApiMapping'

-> GetApiMapping

Create a value of GetApiMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiMappingId:GetApiMapping', getApiMapping_apiMappingId - The API mapping identifier.

$sel:domainName:GetApiMapping', getApiMapping_domainName - The domain name.

Request Lenses

getApiMapping_apiMappingId :: Lens' GetApiMapping Text

The API mapping identifier.

getApiMapping_domainName :: Lens' GetApiMapping Text

The domain name.

Destructuring the Response

data GetApiMappingResponse

See: newGetApiMappingResponse smart constructor.

Constructors

GetApiMappingResponse'

Fields

stage :: Maybe Text

The API stage.

apiId :: Maybe Text

The API identifier.

apiMappingKey :: Maybe Text

The API mapping key.

apiMappingId :: Maybe Text

The API mapping identifier.

httpStatus :: Int

The response's http status code.

newGetApiMappingResponse

Arguments

:: Int
$sel:httpStatus:GetApiMappingResponse'

-> GetApiMappingResponse

Create a value of GetApiMappingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stage:GetApiMappingResponse', getApiMappingResponse_stage - The API stage.

$sel:apiId:GetApiMappingResponse', getApiMappingResponse_apiId - The API identifier.

$sel:apiMappingKey:GetApiMappingResponse', getApiMappingResponse_apiMappingKey - The API mapping key.

$sel:apiMappingId:GetApiMapping', getApiMappingResponse_apiMappingId - The API mapping identifier.

$sel:httpStatus:GetApiMappingResponse', getApiMappingResponse_httpStatus - The response's http status code.

Response Lenses

getApiMappingResponse_stage :: Lens' GetApiMappingResponse (Maybe Text)

The API stage.

getApiMappingResponse_apiId :: Lens' GetApiMappingResponse (Maybe Text)

The API identifier.

getApiMappingResponse_apiMappingKey :: Lens' GetApiMappingResponse (Maybe Text)

The API mapping key.

getApiMappingResponse_apiMappingId :: Lens' GetApiMappingResponse (Maybe Text)

The API mapping identifier.

getApiMappingResponse_httpStatus :: Lens' GetApiMappingResponse Int

The response's http status code.