Network.AWS.ApiGatewayV2.GetIntegration

Description

Gets an Integration.

Creating a Request

data GetIntegration

See: newGetIntegration smart constructor.

Constructors

GetIntegration'

Fields

apiId :: Text

The API identifier.

integrationId :: Text

The integration ID.

newGetIntegration

Arguments

:: Text
$sel:apiId:GetIntegration'

-> Text
$sel:integrationId:GetIntegration'

-> GetIntegration

Create a value of GetIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:GetIntegration', getIntegration_apiId - The API identifier.

$sel:integrationId:GetIntegration', getIntegration_integrationId - The integration ID.

Request Lenses

getIntegration_apiId :: Lens' GetIntegration Text

The API identifier.

getIntegration_integrationId :: Lens' GetIntegration Text

The integration ID.

Destructuring the Response

data GetIntegrationResponse'

See: newGetIntegrationResponse' smart constructor.

Constructors

GetIntegrationResponse''

Fields

integrationResponseSelectionExpression :: Maybe Text

The integration response selection expression for the integration. Supported only for WebSocket APIs. See Integration Response Selection Expressions.

requestTemplates :: Maybe (HashMap Text Text)

Represents a map of Velocity templates that are applied on the request payload based on the value of the Content-Type header sent by the client. The content type value is the key in this map, and the template (as a String) is the value. Supported only for WebSocket APIs.

integrationSubtype :: Maybe Text

Supported only for HTTP API AWS_PROXY integrations. Specifies the AWS service action to invoke. To learn more, see Integration subtype reference.

credentialsArn :: Maybe Text

Specifies the credentials required for the integration, if any. For AWS integrations, three options are available. To specify an IAM Role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To require that the caller's identity be passed through from the request, specify the string arn:aws:iam::*:user/*. To use resource-based permissions on supported AWS services, specify null.

integrationUri :: Maybe Text

For a Lambda integration, specify the URI of a Lambda function.

For an HTTP integration, specify a fully-qualified URL.

For an HTTP API private integration, specify the ARN of an Application Load Balancer listener, Network Load Balancer listener, or AWS Cloud Map service. If you specify the ARN of an AWS Cloud Map service, API Gateway uses DiscoverInstances to identify resources. You can use query parameters to target specific resources. To learn more, see DiscoverInstances. For private integrations, all resources must be owned by the same AWS account.

integrationId :: Maybe Text

Represents the identifier of an integration.

requestParameters :: Maybe (HashMap Text Text)

For WebSocket APIs, a key-value map specifying request parameters that are passed from the method request to the backend. The key is an integration request parameter name and the associated value is a method request parameter value or static value that must be enclosed within single quotes and pre-encoded as required by the backend. The method request parameter value must match the pattern of method.request.{location}.{name} , where {location} is querystring, path, or header; and {name} must be a valid and unique method request parameter name.

For HTTP API integrations with a specified integrationSubtype, request parameters are a key-value map specifying parameters that are passed to AWS_PROXY integrations. You can provide static values, or map request data, stage variables, or context variables that are evaluated at runtime. To learn more, see Working with AWS service integrations for HTTP APIs.

For HTTP API itegrations, without a specified integrationSubtype request parameters are a key-value map specifying how to transform HTTP requests before sending them to backend integrations. The key should follow the pattern <action>:<header|querystring|path>.<location>. The action can be append, overwrite or remove. For values, you can provide static values, or map request data, stage variables, or context variables that are evaluated at runtime. To learn more, see Transforming API requests and responses.

connectionId :: Maybe Text

The ID of the VPC link for a private integration. Supported only for HTTP APIs.

passthroughBehavior :: Maybe PassthroughBehavior

Specifies the pass-through behavior for incoming requests based on the Content-Type header in the request, and the available mapping templates specified as the requestTemplates property on the Integration resource. There are three valid values: WHEN_NO_MATCH, WHEN_NO_TEMPLATES, and NEVER. Supported only for WebSocket APIs.

WHEN_NO_MATCH passes the request body for unmapped content types through to the integration backend without transformation.

NEVER rejects unmapped content types with an HTTP 415 Unsupported Media Type response.

WHEN_NO_TEMPLATES allows pass-through when the integration has no content types mapped to templates. However, if there is at least one content type defined, unmapped content types will be rejected with the same HTTP 415 Unsupported Media Type response.

integrationMethod :: Maybe Text

Specifies the integration's HTTP method type.

tlsConfig :: Maybe TlsConfig

The TLS configuration for a private integration. If you specify a TLS configuration, private integration traffic uses the HTTPS protocol. Supported only for HTTP APIs.

payloadFormatVersion :: Maybe Text

Specifies the format of the payload sent to an integration. Required for HTTP APIs.

templateSelectionExpression :: Maybe Text

The template selection expression for the integration. Supported only for WebSocket APIs.

timeoutInMillis :: Maybe Natural

Custom timeout between 50 and 29,000 milliseconds for WebSocket APIs and between 50 and 30,000 milliseconds for HTTP APIs. The default timeout is 29 seconds for WebSocket APIs and 30 seconds for HTTP APIs.

apiGatewayManaged :: Maybe Bool

Specifies whether an integration is managed by API Gateway. If you created an API using using quick create, the resulting integration is managed by API Gateway. You can update a managed integration, but you can't delete it.

contentHandlingStrategy :: Maybe ContentHandlingStrategy

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

integrationType :: Maybe IntegrationType

The integration type of an integration. One of the following:

AWS: for integrating the route or method request with an AWS service action, including the Lambda function-invoking action. With the Lambda function-invoking action, this is referred to as the Lambda custom integration. With any other AWS service action, this is known as AWS integration. Supported only for WebSocket APIs.

AWS_PROXY: for integrating the route or method request with a Lambda function or other AWS service action. This integration is also referred to as a Lambda proxy integration.

HTTP: for integrating the route or method request with an HTTP endpoint. This integration is also referred to as the HTTP custom integration. Supported only for WebSocket APIs.

HTTP_PROXY: for integrating the route or method request with an HTTP endpoint, with the client request passed through as-is. This is also referred to as HTTP proxy integration.

MOCK: for integrating the route or method request with API Gateway as a "loopback" endpoint without invoking any backend. Supported only for WebSocket APIs.

description :: Maybe Text

Represents the description of an integration.

connectionType :: Maybe ConnectionType

The type of the network connection to the integration endpoint. Specify INTERNET for connections through the public routable internet or VPC_LINK for private connections between API Gateway and resources in a VPC. The default value is INTERNET.

responseParameters :: Maybe (HashMap Text (HashMap Text Text))

Supported only for HTTP APIs. You use response parameters to transform the HTTP response from a backend integration before returning the response to clients. Specify a key-value map from a selection key to response parameters. The selection key must be a valid HTTP status code within the range of 200-599. Response parameters are a key-value map. The key must match pattern <action>:<header>.<location> or overwrite.statuscode. The action can be append, overwrite or remove. The value can be a static value, or map to response data, stage variables, or context variables that are evaluated at runtime. To learn more, see Transforming API requests and responses.

httpStatus :: Int

The response's http status code.

newGetIntegrationResponse'

Arguments

:: Int
$sel:httpStatus:GetIntegrationResponse''

-> GetIntegrationResponse'

Create a value of GetIntegrationResponse' with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:integrationResponseSelectionExpression:GetIntegrationResponse'', getIntegrationResponse'_integrationResponseSelectionExpression - The integration response selection expression for the integration. Supported only for WebSocket APIs. See Integration Response Selection Expressions.

$sel:requestTemplates:GetIntegrationResponse'', getIntegrationResponse'_requestTemplates - Represents a map of Velocity templates that are applied on the request payload based on the value of the Content-Type header sent by the client. The content type value is the key in this map, and the template (as a String) is the value. Supported only for WebSocket APIs.

$sel:integrationSubtype:GetIntegrationResponse'', getIntegrationResponse'_integrationSubtype - Supported only for HTTP API AWS_PROXY integrations. Specifies the AWS service action to invoke. To learn more, see Integration subtype reference.

$sel:credentialsArn:GetIntegrationResponse'', getIntegrationResponse'_credentialsArn - Specifies the credentials required for the integration, if any. For AWS integrations, three options are available. To specify an IAM Role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To require that the caller's identity be passed through from the request, specify the string arn:aws:iam::*:user/*. To use resource-based permissions on supported AWS services, specify null.

$sel:integrationUri:GetIntegrationResponse'', getIntegrationResponse'_integrationUri - For a Lambda integration, specify the URI of a Lambda function.

For an HTTP integration, specify a fully-qualified URL.

For an HTTP API private integration, specify the ARN of an Application Load Balancer listener, Network Load Balancer listener, or AWS Cloud Map service. If you specify the ARN of an AWS Cloud Map service, API Gateway uses DiscoverInstances to identify resources. You can use query parameters to target specific resources. To learn more, see DiscoverInstances. For private integrations, all resources must be owned by the same AWS account.

$sel:integrationId:GetIntegration', getIntegrationResponse'_integrationId - Represents the identifier of an integration.

$sel:requestParameters:GetIntegrationResponse'', getIntegrationResponse'_requestParameters - For WebSocket APIs, a key-value map specifying request parameters that are passed from the method request to the backend. The key is an integration request parameter name and the associated value is a method request parameter value or static value that must be enclosed within single quotes and pre-encoded as required by the backend. The method request parameter value must match the pattern of method.request.{location}.{name} , where {location} is querystring, path, or header; and {name} must be a valid and unique method request parameter name.

For HTTP API integrations with a specified integrationSubtype, request parameters are a key-value map specifying parameters that are passed to AWS_PROXY integrations. You can provide static values, or map request data, stage variables, or context variables that are evaluated at runtime. To learn more, see Working with AWS service integrations for HTTP APIs.

For HTTP API itegrations, without a specified integrationSubtype request parameters are a key-value map specifying how to transform HTTP requests before sending them to backend integrations. The key should follow the pattern <action>:<header|querystring|path>.<location>. The action can be append, overwrite or remove. For values, you can provide static values, or map request data, stage variables, or context variables that are evaluated at runtime. To learn more, see Transforming API requests and responses.

$sel:connectionId:GetIntegrationResponse'', getIntegrationResponse'_connectionId - The ID of the VPC link for a private integration. Supported only for HTTP APIs.

$sel:passthroughBehavior:GetIntegrationResponse'', getIntegrationResponse'_passthroughBehavior - Specifies the pass-through behavior for incoming requests based on the Content-Type header in the request, and the available mapping templates specified as the requestTemplates property on the Integration resource. There are three valid values: WHEN_NO_MATCH, WHEN_NO_TEMPLATES, and NEVER. Supported only for WebSocket APIs.

WHEN_NO_MATCH passes the request body for unmapped content types through to the integration backend without transformation.

NEVER rejects unmapped content types with an HTTP 415 Unsupported Media Type response.

WHEN_NO_TEMPLATES allows pass-through when the integration has no content types mapped to templates. However, if there is at least one content type defined, unmapped content types will be rejected with the same HTTP 415 Unsupported Media Type response.

$sel:integrationMethod:GetIntegrationResponse'', getIntegrationResponse'_integrationMethod - Specifies the integration's HTTP method type.

$sel:tlsConfig:GetIntegrationResponse'', getIntegrationResponse'_tlsConfig - The TLS configuration for a private integration. If you specify a TLS configuration, private integration traffic uses the HTTPS protocol. Supported only for HTTP APIs.

$sel:payloadFormatVersion:GetIntegrationResponse'', getIntegrationResponse'_payloadFormatVersion - Specifies the format of the payload sent to an integration. Required for HTTP APIs.

$sel:templateSelectionExpression:GetIntegrationResponse'', getIntegrationResponse'_templateSelectionExpression - The template selection expression for the integration. Supported only for WebSocket APIs.

$sel:timeoutInMillis:GetIntegrationResponse'', getIntegrationResponse'_timeoutInMillis - Custom timeout between 50 and 29,000 milliseconds for WebSocket APIs and between 50 and 30,000 milliseconds for HTTP APIs. The default timeout is 29 seconds for WebSocket APIs and 30 seconds for HTTP APIs.

$sel:apiGatewayManaged:GetIntegrationResponse'', getIntegrationResponse'_apiGatewayManaged - Specifies whether an integration is managed by API Gateway. If you created an API using using quick create, the resulting integration is managed by API Gateway. You can update a managed integration, but you can't delete it.

$sel:contentHandlingStrategy:GetIntegrationResponse'', getIntegrationResponse'_contentHandlingStrategy - Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

$sel:integrationType:GetIntegrationResponse'', getIntegrationResponse'_integrationType - The integration type of an integration. One of the following:

AWS: for integrating the route or method request with an AWS service action, including the Lambda function-invoking action. With the Lambda function-invoking action, this is referred to as the Lambda custom integration. With any other AWS service action, this is known as AWS integration. Supported only for WebSocket APIs.

AWS_PROXY: for integrating the route or method request with a Lambda function or other AWS service action. This integration is also referred to as a Lambda proxy integration.

HTTP: for integrating the route or method request with an HTTP endpoint. This integration is also referred to as the HTTP custom integration. Supported only for WebSocket APIs.

HTTP_PROXY: for integrating the route or method request with an HTTP endpoint, with the client request passed through as-is. This is also referred to as HTTP proxy integration.

MOCK: for integrating the route or method request with API Gateway as a "loopback" endpoint without invoking any backend. Supported only for WebSocket APIs.

$sel:description:GetIntegrationResponse'', getIntegrationResponse'_description - Represents the description of an integration.

$sel:connectionType:GetIntegrationResponse'', getIntegrationResponse'_connectionType - The type of the network connection to the integration endpoint. Specify INTERNET for connections through the public routable internet or VPC_LINK for private connections between API Gateway and resources in a VPC. The default value is INTERNET.

$sel:responseParameters:GetIntegrationResponse'', getIntegrationResponse'_responseParameters - Supported only for HTTP APIs. You use response parameters to transform the HTTP response from a backend integration before returning the response to clients. Specify a key-value map from a selection key to response parameters. The selection key must be a valid HTTP status code within the range of 200-599. Response parameters are a key-value map. The key must match pattern <action>:<header>.<location> or overwrite.statuscode. The action can be append, overwrite or remove. The value can be a static value, or map to response data, stage variables, or context variables that are evaluated at runtime. To learn more, see Transforming API requests and responses.

$sel:httpStatus:GetIntegrationResponse'', getIntegrationResponse'_httpStatus - The response's http status code.

Response Lenses

getIntegrationResponse'_integrationResponseSelectionExpression :: Lens' GetIntegrationResponse' (Maybe Text)

The integration response selection expression for the integration. Supported only for WebSocket APIs. See Integration Response Selection Expressions.

getIntegrationResponse'_requestTemplates :: Lens' GetIntegrationResponse' (Maybe (HashMap Text Text))

Represents a map of Velocity templates that are applied on the request payload based on the value of the Content-Type header sent by the client. The content type value is the key in this map, and the template (as a String) is the value. Supported only for WebSocket APIs.

getIntegrationResponse'_integrationSubtype :: Lens' GetIntegrationResponse' (Maybe Text)

Supported only for HTTP API AWS_PROXY integrations. Specifies the AWS service action to invoke. To learn more, see Integration subtype reference.

getIntegrationResponse'_credentialsArn :: Lens' GetIntegrationResponse' (Maybe Text)

Specifies the credentials required for the integration, if any. For AWS integrations, three options are available. To specify an IAM Role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To require that the caller's identity be passed through from the request, specify the string arn:aws:iam::*:user/*. To use resource-based permissions on supported AWS services, specify null.

getIntegrationResponse'_integrationUri :: Lens' GetIntegrationResponse' (Maybe Text)

For a Lambda integration, specify the URI of a Lambda function.

For an HTTP integration, specify a fully-qualified URL.

For an HTTP API private integration, specify the ARN of an Application Load Balancer listener, Network Load Balancer listener, or AWS Cloud Map service. If you specify the ARN of an AWS Cloud Map service, API Gateway uses DiscoverInstances to identify resources. You can use query parameters to target specific resources. To learn more, see DiscoverInstances. For private integrations, all resources must be owned by the same AWS account.

getIntegrationResponse'_integrationId :: Lens' GetIntegrationResponse' (Maybe Text)

Represents the identifier of an integration.

getIntegrationResponse'_requestParameters :: Lens' GetIntegrationResponse' (Maybe (HashMap Text Text))

For WebSocket APIs, a key-value map specifying request parameters that are passed from the method request to the backend. The key is an integration request parameter name and the associated value is a method request parameter value or static value that must be enclosed within single quotes and pre-encoded as required by the backend. The method request parameter value must match the pattern of method.request.{location}.{name} , where {location} is querystring, path, or header; and {name} must be a valid and unique method request parameter name.

For HTTP API integrations with a specified integrationSubtype, request parameters are a key-value map specifying parameters that are passed to AWS_PROXY integrations. You can provide static values, or map request data, stage variables, or context variables that are evaluated at runtime. To learn more, see Working with AWS service integrations for HTTP APIs.

For HTTP API itegrations, without a specified integrationSubtype request parameters are a key-value map specifying how to transform HTTP requests before sending them to backend integrations. The key should follow the pattern <action>:<header|querystring|path>.<location>. The action can be append, overwrite or remove. For values, you can provide static values, or map request data, stage variables, or context variables that are evaluated at runtime. To learn more, see Transforming API requests and responses.

getIntegrationResponse'_connectionId :: Lens' GetIntegrationResponse' (Maybe Text)

The ID of the VPC link for a private integration. Supported only for HTTP APIs.

getIntegrationResponse'_passthroughBehavior :: Lens' GetIntegrationResponse' (Maybe PassthroughBehavior)

Specifies the pass-through behavior for incoming requests based on the Content-Type header in the request, and the available mapping templates specified as the requestTemplates property on the Integration resource. There are three valid values: WHEN_NO_MATCH, WHEN_NO_TEMPLATES, and NEVER. Supported only for WebSocket APIs.

WHEN_NO_MATCH passes the request body for unmapped content types through to the integration backend without transformation.

NEVER rejects unmapped content types with an HTTP 415 Unsupported Media Type response.

WHEN_NO_TEMPLATES allows pass-through when the integration has no content types mapped to templates. However, if there is at least one content type defined, unmapped content types will be rejected with the same HTTP 415 Unsupported Media Type response.

getIntegrationResponse'_integrationMethod :: Lens' GetIntegrationResponse' (Maybe Text)

Specifies the integration's HTTP method type.

getIntegrationResponse'_tlsConfig :: Lens' GetIntegrationResponse' (Maybe TlsConfig)

The TLS configuration for a private integration. If you specify a TLS configuration, private integration traffic uses the HTTPS protocol. Supported only for HTTP APIs.

getIntegrationResponse'_payloadFormatVersion :: Lens' GetIntegrationResponse' (Maybe Text)

Specifies the format of the payload sent to an integration. Required for HTTP APIs.

getIntegrationResponse'_templateSelectionExpression :: Lens' GetIntegrationResponse' (Maybe Text)

The template selection expression for the integration. Supported only for WebSocket APIs.

getIntegrationResponse'_timeoutInMillis :: Lens' GetIntegrationResponse' (Maybe Natural)

Custom timeout between 50 and 29,000 milliseconds for WebSocket APIs and between 50 and 30,000 milliseconds for HTTP APIs. The default timeout is 29 seconds for WebSocket APIs and 30 seconds for HTTP APIs.

getIntegrationResponse'_apiGatewayManaged :: Lens' GetIntegrationResponse' (Maybe Bool)

Specifies whether an integration is managed by API Gateway. If you created an API using using quick create, the resulting integration is managed by API Gateway. You can update a managed integration, but you can't delete it.

getIntegrationResponse'_contentHandlingStrategy :: Lens' GetIntegrationResponse' (Maybe ContentHandlingStrategy)

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

getIntegrationResponse'_integrationType :: Lens' GetIntegrationResponse' (Maybe IntegrationType)

The integration type of an integration. One of the following:

AWS: for integrating the route or method request with an AWS service action, including the Lambda function-invoking action. With the Lambda function-invoking action, this is referred to as the Lambda custom integration. With any other AWS service action, this is known as AWS integration. Supported only for WebSocket APIs.

AWS_PROXY: for integrating the route or method request with a Lambda function or other AWS service action. This integration is also referred to as a Lambda proxy integration.

HTTP: for integrating the route or method request with an HTTP endpoint. This integration is also referred to as the HTTP custom integration. Supported only for WebSocket APIs.

HTTP_PROXY: for integrating the route or method request with an HTTP endpoint, with the client request passed through as-is. This is also referred to as HTTP proxy integration.

MOCK: for integrating the route or method request with API Gateway as a "loopback" endpoint without invoking any backend. Supported only for WebSocket APIs.

getIntegrationResponse'_description :: Lens' GetIntegrationResponse' (Maybe Text)

Represents the description of an integration.

getIntegrationResponse'_connectionType :: Lens' GetIntegrationResponse' (Maybe ConnectionType)

The type of the network connection to the integration endpoint. Specify INTERNET for connections through the public routable internet or VPC_LINK for private connections between API Gateway and resources in a VPC. The default value is INTERNET.

getIntegrationResponse'_responseParameters :: Lens' GetIntegrationResponse' (Maybe (HashMap Text (HashMap Text Text)))

Supported only for HTTP APIs. You use response parameters to transform the HTTP response from a backend integration before returning the response to clients. Specify a key-value map from a selection key to response parameters. The selection key must be a valid HTTP status code within the range of 200-599. Response parameters are a key-value map. The key must match pattern <action>:<header>.<location> or overwrite.statuscode. The action can be append, overwrite or remove. The value can be a static value, or map to response data, stage variables, or context variables that are evaluated at runtime. To learn more, see Transforming API requests and responses.

getIntegrationResponse'_httpStatus :: Lens' GetIntegrationResponse' Int

The response's http status code.