Network.AWS.ApiGatewayV2.GetIntegrationResponses

Description

Gets the IntegrationResponses for an Integration.

This operation returns paginated results.

Creating a Request

data GetIntegrationResponses

See: newGetIntegrationResponses smart constructor.

Constructors

GetIntegrationResponses'

Fields

nextToken :: Maybe Text

The next page of elements from this collection. Not valid for the last element of the collection.

maxResults :: Maybe Text

The maximum number of elements to be returned for this resource.

integrationId :: Text

The integration ID.

apiId :: Text

The API identifier.

newGetIntegrationResponses

Arguments

:: Text
$sel:integrationId:GetIntegrationResponses'

-> Text
$sel:apiId:GetIntegrationResponses'

-> GetIntegrationResponses

Create a value of GetIntegrationResponses with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetIntegrationResponses', getIntegrationResponses_nextToken - The next page of elements from this collection. Not valid for the last element of the collection.

$sel:maxResults:GetIntegrationResponses', getIntegrationResponses_maxResults - The maximum number of elements to be returned for this resource.

$sel:integrationId:GetIntegrationResponses', getIntegrationResponses_integrationId - The integration ID.

$sel:apiId:GetIntegrationResponses', getIntegrationResponses_apiId - The API identifier.

Request Lenses

getIntegrationResponses_nextToken :: Lens' GetIntegrationResponses (Maybe Text)

The next page of elements from this collection. Not valid for the last element of the collection.

getIntegrationResponses_maxResults :: Lens' GetIntegrationResponses (Maybe Text)

The maximum number of elements to be returned for this resource.

getIntegrationResponses_integrationId :: Lens' GetIntegrationResponses Text

The integration ID.

getIntegrationResponses_apiId :: Lens' GetIntegrationResponses Text

The API identifier.

Destructuring the Response

data GetIntegrationResponsesResponse

See: newGetIntegrationResponsesResponse smart constructor.

Constructors

GetIntegrationResponsesResponse'

Fields

items :: Maybe [IntegrationResponse]

The elements from this collection.

nextToken :: Maybe Text

The next page of elements from this collection. Not valid for the last element of the collection.

httpStatus :: Int

The response's http status code.

newGetIntegrationResponsesResponse

Arguments

:: Int
$sel:httpStatus:GetIntegrationResponsesResponse'

-> GetIntegrationResponsesResponse

Create a value of GetIntegrationResponsesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:GetIntegrationResponsesResponse', getIntegrationResponsesResponse_items - The elements from this collection.

$sel:nextToken:GetIntegrationResponses', getIntegrationResponsesResponse_nextToken - The next page of elements from this collection. Not valid for the last element of the collection.

$sel:httpStatus:GetIntegrationResponsesResponse', getIntegrationResponsesResponse_httpStatus - The response's http status code.

Response Lenses

getIntegrationResponsesResponse_items :: Lens' GetIntegrationResponsesResponse (Maybe [IntegrationResponse])

The elements from this collection.

getIntegrationResponsesResponse_nextToken :: Lens' GetIntegrationResponsesResponse (Maybe Text)

The next page of elements from this collection. Not valid for the last element of the collection.

getIntegrationResponsesResponse_httpStatus :: Lens' GetIntegrationResponsesResponse Int

The response's http status code.