Network.AWS.ApiGatewayV2.GetRoute

Description

Gets a Route.

Creating a Request

data GetRoute

See: newGetRoute smart constructor.

Constructors

GetRoute'

Fields

apiId :: Text

The API identifier.

routeId :: Text

The route ID.

newGetRoute

Arguments

:: Text
$sel:apiId:GetRoute'

-> Text
$sel:routeId:GetRoute'

-> GetRoute

Create a value of GetRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:GetRoute', getRoute_apiId - The API identifier.

$sel:routeId:GetRoute', getRoute_routeId - The route ID.

Request Lenses

getRoute_apiId :: Lens' GetRoute Text

The API identifier.

getRoute_routeId :: Lens' GetRoute Text

The route ID.

Destructuring the Response

data GetRouteResponse'

See: newGetRouteResponse' smart constructor.

Constructors

GetRouteResponse''

Fields

authorizationScopes :: Maybe [Text]

A list of authorization scopes configured on a route. The scopes are used with a JWT authorizer to authorize the method invocation. The authorization works by matching the route scopes against the scopes parsed from the access token in the incoming request. The method invocation is authorized if any route scope matches a claimed scope in the access token. Otherwise, the invocation is not authorized. When the route scope is configured, the client must provide an access token instead of an identity token for authorization purposes.

modelSelectionExpression :: Maybe Text

The model selection expression for the route. Supported only for WebSocket APIs.

requestModels :: Maybe (HashMap Text Text)

The request models for the route. Supported only for WebSocket APIs.

routeResponseSelectionExpression :: Maybe Text

The route response selection expression for the route. Supported only for WebSocket APIs.

requestParameters :: Maybe (HashMap Text ParameterConstraints)

The request parameters for the route. Supported only for WebSocket APIs.

routeId :: Maybe Text

The route ID.

authorizerId :: Maybe Text

The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

operationName :: Maybe Text

The operation name for the route.

apiGatewayManaged :: Maybe Bool

Specifies whether a route is managed by API Gateway. If you created an API using quick create, the $default route is managed by API Gateway. You can't modify the $default route key.

authorizationType :: Maybe AuthorizationType

The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

apiKeyRequired :: Maybe Bool

Specifies whether an API key is required for this route. Supported only for WebSocket APIs.

routeKey :: Maybe Text

The route key for the route.

target :: Maybe Text

The target for the route.

httpStatus :: Int

The response's http status code.

newGetRouteResponse'

Arguments

:: Int
$sel:httpStatus:GetRouteResponse''

-> GetRouteResponse'

Create a value of GetRouteResponse' with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizationScopes:GetRouteResponse'', getRouteResponse'_authorizationScopes - A list of authorization scopes configured on a route. The scopes are used with a JWT authorizer to authorize the method invocation. The authorization works by matching the route scopes against the scopes parsed from the access token in the incoming request. The method invocation is authorized if any route scope matches a claimed scope in the access token. Otherwise, the invocation is not authorized. When the route scope is configured, the client must provide an access token instead of an identity token for authorization purposes.

$sel:modelSelectionExpression:GetRouteResponse'', getRouteResponse'_modelSelectionExpression - The model selection expression for the route. Supported only for WebSocket APIs.

$sel:requestModels:GetRouteResponse'', getRouteResponse'_requestModels - The request models for the route. Supported only for WebSocket APIs.

$sel:routeResponseSelectionExpression:GetRouteResponse'', getRouteResponse'_routeResponseSelectionExpression - The route response selection expression for the route. Supported only for WebSocket APIs.

$sel:requestParameters:GetRouteResponse'', getRouteResponse'_requestParameters - The request parameters for the route. Supported only for WebSocket APIs.

$sel:routeId:GetRoute', getRouteResponse'_routeId - The route ID.

$sel:authorizerId:GetRouteResponse'', getRouteResponse'_authorizerId - The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

$sel:operationName:GetRouteResponse'', getRouteResponse'_operationName - The operation name for the route.

$sel:apiGatewayManaged:GetRouteResponse'', getRouteResponse'_apiGatewayManaged - Specifies whether a route is managed by API Gateway. If you created an API using quick create, the $default route is managed by API Gateway. You can't modify the $default route key.

$sel:authorizationType:GetRouteResponse'', getRouteResponse'_authorizationType - The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

$sel:apiKeyRequired:GetRouteResponse'', getRouteResponse'_apiKeyRequired - Specifies whether an API key is required for this route. Supported only for WebSocket APIs.

$sel:routeKey:GetRouteResponse'', getRouteResponse'_routeKey - The route key for the route.

$sel:target:GetRouteResponse'', getRouteResponse'_target - The target for the route.

$sel:httpStatus:GetRouteResponse'', getRouteResponse'_httpStatus - The response's http status code.

Response Lenses

getRouteResponse'_authorizationScopes :: Lens' GetRouteResponse' (Maybe [Text])

A list of authorization scopes configured on a route. The scopes are used with a JWT authorizer to authorize the method invocation. The authorization works by matching the route scopes against the scopes parsed from the access token in the incoming request. The method invocation is authorized if any route scope matches a claimed scope in the access token. Otherwise, the invocation is not authorized. When the route scope is configured, the client must provide an access token instead of an identity token for authorization purposes.

getRouteResponse'_modelSelectionExpression :: Lens' GetRouteResponse' (Maybe Text)

The model selection expression for the route. Supported only for WebSocket APIs.

getRouteResponse'_requestModels :: Lens' GetRouteResponse' (Maybe (HashMap Text Text))

The request models for the route. Supported only for WebSocket APIs.

getRouteResponse'_routeResponseSelectionExpression :: Lens' GetRouteResponse' (Maybe Text)

The route response selection expression for the route. Supported only for WebSocket APIs.

getRouteResponse'_requestParameters :: Lens' GetRouteResponse' (Maybe (HashMap Text ParameterConstraints))

The request parameters for the route. Supported only for WebSocket APIs.

getRouteResponse'_routeId :: Lens' GetRouteResponse' (Maybe Text)

The route ID.

getRouteResponse'_authorizerId :: Lens' GetRouteResponse' (Maybe Text)

The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

getRouteResponse'_operationName :: Lens' GetRouteResponse' (Maybe Text)

The operation name for the route.

getRouteResponse'_apiGatewayManaged :: Lens' GetRouteResponse' (Maybe Bool)

Specifies whether a route is managed by API Gateway. If you created an API using quick create, the $default route is managed by API Gateway. You can't modify the $default route key.

getRouteResponse'_authorizationType :: Lens' GetRouteResponse' (Maybe AuthorizationType)

The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWS_IAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

getRouteResponse'_apiKeyRequired :: Lens' GetRouteResponse' (Maybe Bool)

Specifies whether an API key is required for this route. Supported only for WebSocket APIs.

getRouteResponse'_routeKey :: Lens' GetRouteResponse' (Maybe Text)

The route key for the route.

getRouteResponse'_target :: Lens' GetRouteResponse' (Maybe Text)

The target for the route.

getRouteResponse'_httpStatus :: Lens' GetRouteResponse' Int

The response's http status code.