Network.AWS.ApiGatewayV2.GetStage

Description

Gets a Stage.

Creating a Request

data GetStage

See: newGetStage smart constructor.

Constructors

GetStage'

Fields

stageName :: Text

The stage name. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

apiId :: Text

The API identifier.

newGetStage

Arguments

:: Text
$sel:stageName:GetStage'

-> Text
$sel:apiId:GetStage'

-> GetStage

Create a value of GetStage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stageName:GetStage', getStage_stageName - The stage name. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

$sel:apiId:GetStage', getStage_apiId - The API identifier.

Request Lenses

getStage_stageName :: Lens' GetStage Text

The stage name. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

getStage_apiId :: Lens' GetStage Text

The API identifier.

Destructuring the Response

data GetStageResponse

See: newGetStageResponse smart constructor.

Constructors

GetStageResponse'

Fields

lastDeploymentStatusMessage :: Maybe Text

Describes the status of the last deployment of a stage. Supported only for stages with autoDeploy enabled.

deploymentId :: Maybe Text

The identifier of the Deployment that the Stage is associated with. Can't be updated if autoDeploy is enabled.

routeSettings :: Maybe (HashMap Text RouteSettings)

Route settings for the stage, by routeKey.

accessLogSettings :: Maybe AccessLogSettings

Settings for logging access in this stage.

clientCertificateId :: Maybe Text

The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

stageVariables :: Maybe (HashMap Text Text)

A map that defines the stage variables for a stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

autoDeploy :: Maybe Bool

Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

createdDate :: Maybe POSIX

The timestamp when the stage was created.

defaultRouteSettings :: Maybe RouteSettings

Default route settings for the stage.

apiGatewayManaged :: Maybe Bool

Specifies whether a stage is managed by API Gateway. If you created an API using quick create, the $default stage is managed by API Gateway. You can't modify the $default stage.

stageName :: Maybe Text

The name of the stage.

lastUpdatedDate :: Maybe POSIX

The timestamp when the stage was last updated.

description :: Maybe Text

The description of the stage.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

httpStatus :: Int

The response's http status code.

newGetStageResponse

Arguments

:: Int
$sel:httpStatus:GetStageResponse'

-> GetStageResponse

Create a value of GetStageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastDeploymentStatusMessage:GetStageResponse', getStageResponse_lastDeploymentStatusMessage - Describes the status of the last deployment of a stage. Supported only for stages with autoDeploy enabled.

$sel:deploymentId:GetStageResponse', getStageResponse_deploymentId - The identifier of the Deployment that the Stage is associated with. Can't be updated if autoDeploy is enabled.

$sel:routeSettings:GetStageResponse', getStageResponse_routeSettings - Route settings for the stage, by routeKey.

$sel:accessLogSettings:GetStageResponse', getStageResponse_accessLogSettings - Settings for logging access in this stage.

$sel:clientCertificateId:GetStageResponse', getStageResponse_clientCertificateId - The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

$sel:stageVariables:GetStageResponse', getStageResponse_stageVariables - A map that defines the stage variables for a stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

$sel:autoDeploy:GetStageResponse', getStageResponse_autoDeploy - Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

$sel:createdDate:GetStageResponse', getStageResponse_createdDate - The timestamp when the stage was created.

$sel:defaultRouteSettings:GetStageResponse', getStageResponse_defaultRouteSettings - Default route settings for the stage.

$sel:apiGatewayManaged:GetStageResponse', getStageResponse_apiGatewayManaged - Specifies whether a stage is managed by API Gateway. If you created an API using quick create, the $default stage is managed by API Gateway. You can't modify the $default stage.

$sel:stageName:GetStage', getStageResponse_stageName - The name of the stage.

$sel:lastUpdatedDate:GetStageResponse', getStageResponse_lastUpdatedDate - The timestamp when the stage was last updated.

$sel:description:GetStageResponse', getStageResponse_description - The description of the stage.

$sel:tags:GetStageResponse', getStageResponse_tags - The collection of tags. Each tag element is associated with a given resource.

$sel:httpStatus:GetStageResponse', getStageResponse_httpStatus - The response's http status code.

Response Lenses

getStageResponse_lastDeploymentStatusMessage :: Lens' GetStageResponse (Maybe Text)

Describes the status of the last deployment of a stage. Supported only for stages with autoDeploy enabled.

getStageResponse_deploymentId :: Lens' GetStageResponse (Maybe Text)

The identifier of the Deployment that the Stage is associated with. Can't be updated if autoDeploy is enabled.

getStageResponse_routeSettings :: Lens' GetStageResponse (Maybe (HashMap Text RouteSettings))

Route settings for the stage, by routeKey.

getStageResponse_accessLogSettings :: Lens' GetStageResponse (Maybe AccessLogSettings)

Settings for logging access in this stage.

getStageResponse_clientCertificateId :: Lens' GetStageResponse (Maybe Text)

The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

getStageResponse_stageVariables :: Lens' GetStageResponse (Maybe (HashMap Text Text))

A map that defines the stage variables for a stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

getStageResponse_autoDeploy :: Lens' GetStageResponse (Maybe Bool)

Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

getStageResponse_createdDate :: Lens' GetStageResponse (Maybe UTCTime)

The timestamp when the stage was created.

getStageResponse_defaultRouteSettings :: Lens' GetStageResponse (Maybe RouteSettings)

Default route settings for the stage.

getStageResponse_apiGatewayManaged :: Lens' GetStageResponse (Maybe Bool)

Specifies whether a stage is managed by API Gateway. If you created an API using quick create, the $default stage is managed by API Gateway. You can't modify the $default stage.

getStageResponse_stageName :: Lens' GetStageResponse (Maybe Text)

The name of the stage.

getStageResponse_lastUpdatedDate :: Lens' GetStageResponse (Maybe UTCTime)

The timestamp when the stage was last updated.

getStageResponse_description :: Lens' GetStageResponse (Maybe Text)

The description of the stage.

getStageResponse_tags :: Lens' GetStageResponse (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.

getStageResponse_httpStatus :: Lens' GetStageResponse Int

The response's http status code.