Network.AWS.ApiGatewayV2.GetVpcLink

Description

Gets a VPC link.

Creating a Request

data GetVpcLink

See: newGetVpcLink smart constructor.

Constructors

GetVpcLink'

Fields

vpcLinkId :: Text

The ID of the VPC link.

newGetVpcLink

Arguments

:: Text
$sel:vpcLinkId:GetVpcLink'

-> GetVpcLink

Create a value of GetVpcLink with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcLinkId:GetVpcLink', getVpcLink_vpcLinkId - The ID of the VPC link.

Request Lenses

getVpcLink_vpcLinkId :: Lens' GetVpcLink Text

The ID of the VPC link.

Destructuring the Response

data GetVpcLinkResponse

See: newGetVpcLinkResponse smart constructor.

Constructors

GetVpcLinkResponse'

Fields

securityGroupIds :: Maybe [Text]

A list of security group IDs for the VPC link.

subnetIds :: Maybe [Text]

A list of subnet IDs to include in the VPC link.

vpcLinkId :: Maybe Text

The ID of the VPC link.

createdDate :: Maybe POSIX

The timestamp when the VPC link was created.

vpcLinkVersion :: Maybe VpcLinkVersion

The version of the VPC link.

name :: Maybe Text

The name of the VPC link.

vpcLinkStatusMessage :: Maybe Text

A message summarizing the cause of the status of the VPC link.

tags :: Maybe (HashMap Text Text)

Tags for the VPC link.

vpcLinkStatus :: Maybe VpcLinkStatus

The status of the VPC link.

httpStatus :: Int

The response's http status code.

newGetVpcLinkResponse

Arguments

:: Int
$sel:httpStatus:GetVpcLinkResponse'

-> GetVpcLinkResponse

Create a value of GetVpcLinkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:GetVpcLinkResponse', getVpcLinkResponse_securityGroupIds - A list of security group IDs for the VPC link.

$sel:subnetIds:GetVpcLinkResponse', getVpcLinkResponse_subnetIds - A list of subnet IDs to include in the VPC link.

$sel:vpcLinkId:GetVpcLink', getVpcLinkResponse_vpcLinkId - The ID of the VPC link.

$sel:createdDate:GetVpcLinkResponse', getVpcLinkResponse_createdDate - The timestamp when the VPC link was created.

$sel:vpcLinkVersion:GetVpcLinkResponse', getVpcLinkResponse_vpcLinkVersion - The version of the VPC link.

$sel:name:GetVpcLinkResponse', getVpcLinkResponse_name - The name of the VPC link.

$sel:vpcLinkStatusMessage:GetVpcLinkResponse', getVpcLinkResponse_vpcLinkStatusMessage - A message summarizing the cause of the status of the VPC link.

$sel:tags:GetVpcLinkResponse', getVpcLinkResponse_tags - Tags for the VPC link.

$sel:vpcLinkStatus:GetVpcLinkResponse', getVpcLinkResponse_vpcLinkStatus - The status of the VPC link.

$sel:httpStatus:GetVpcLinkResponse', getVpcLinkResponse_httpStatus - The response's http status code.

Response Lenses

getVpcLinkResponse_securityGroupIds :: Lens' GetVpcLinkResponse (Maybe [Text])

A list of security group IDs for the VPC link.

getVpcLinkResponse_subnetIds :: Lens' GetVpcLinkResponse (Maybe [Text])

A list of subnet IDs to include in the VPC link.

getVpcLinkResponse_vpcLinkId :: Lens' GetVpcLinkResponse (Maybe Text)

The ID of the VPC link.

getVpcLinkResponse_createdDate :: Lens' GetVpcLinkResponse (Maybe UTCTime)

The timestamp when the VPC link was created.

getVpcLinkResponse_vpcLinkVersion :: Lens' GetVpcLinkResponse (Maybe VpcLinkVersion)

The version of the VPC link.

getVpcLinkResponse_name :: Lens' GetVpcLinkResponse (Maybe Text)

The name of the VPC link.

getVpcLinkResponse_vpcLinkStatusMessage :: Lens' GetVpcLinkResponse (Maybe Text)

A message summarizing the cause of the status of the VPC link.

getVpcLinkResponse_tags :: Lens' GetVpcLinkResponse (Maybe (HashMap Text Text))

Tags for the VPC link.

getVpcLinkResponse_vpcLinkStatus :: Lens' GetVpcLinkResponse (Maybe VpcLinkStatus)

The status of the VPC link.

getVpcLinkResponse_httpStatus :: Lens' GetVpcLinkResponse Int

The response's http status code.