Network.AWS.ApiGatewayV2.ImportApi

Description

Imports an API.

Creating a Request

data ImportApi

See: newImportApi smart constructor.

Constructors

ImportApi'

Fields

basepath :: Maybe Text

Specifies how to interpret the base path of the API during import. Valid values are ignore, prepend, and split. The default value is ignore. To learn more, see Set the OpenAPI basePath Property. Supported only for HTTP APIs.

failOnWarnings :: Maybe Bool

Specifies whether to rollback the API creation when a warning is encountered. By default, API creation continues if a warning is encountered.

body :: Text

The OpenAPI definition. Supported only for HTTP APIs.

newImportApi

Arguments

:: Text
$sel:body:ImportApi'

-> ImportApi

Create a value of ImportApi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:basepath:ImportApi', importApi_basepath - Specifies how to interpret the base path of the API during import. Valid values are ignore, prepend, and split. The default value is ignore. To learn more, see Set the OpenAPI basePath Property. Supported only for HTTP APIs.

$sel:failOnWarnings:ImportApi', importApi_failOnWarnings - Specifies whether to rollback the API creation when a warning is encountered. By default, API creation continues if a warning is encountered.

$sel:body:ImportApi', importApi_body - The OpenAPI definition. Supported only for HTTP APIs.

Request Lenses

importApi_basepath :: Lens' ImportApi (Maybe Text)

Specifies how to interpret the base path of the API during import. Valid values are ignore, prepend, and split. The default value is ignore. To learn more, see Set the OpenAPI basePath Property. Supported only for HTTP APIs.

importApi_failOnWarnings :: Lens' ImportApi (Maybe Bool)

Specifies whether to rollback the API creation when a warning is encountered. By default, API creation continues if a warning is encountered.

importApi_body :: Lens' ImportApi Text

The OpenAPI definition. Supported only for HTTP APIs.

Destructuring the Response

data ImportApiResponse

See: newImportApiResponse smart constructor.

Constructors

ImportApiResponse'

Fields

apiId :: Maybe Text

The API ID.

disableExecuteApiEndpoint :: Maybe Bool

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

apiEndpoint :: Maybe Text

The URI of the API, of the form {api-id}.execute-api.{region}.amazonaws.com. The stage name is typically appended to this URI to form a complete path to a deployed API stage.

warnings :: Maybe [Text]

The warning messages reported when failonwarnings is turned on during API import.

createdDate :: Maybe POSIX

The timestamp when the API was created.

name :: Maybe Text

The name of the API.

version :: Maybe Text

A version identifier for the API.

apiGatewayManaged :: Maybe Bool

Specifies whether an API is managed by API Gateway. You can't update or delete a managed API by using API Gateway. A managed API can be deleted only through the tooling or service that created it.

apiKeySelectionExpression :: Maybe Text

An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

corsConfiguration :: Maybe Cors

A CORS configuration. Supported only for HTTP APIs.

routeSelectionExpression :: Maybe Text

The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

importInfo :: Maybe [Text]

The validation information during API import. This may include particular properties of your OpenAPI definition which are ignored during import. Supported only for HTTP APIs.

disableSchemaValidation :: Maybe Bool

Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

description :: Maybe Text

The description of the API.

protocolType :: Maybe ProtocolType

The API protocol.

tags :: Maybe (HashMap Text Text)

A collection of tags associated with the API.

httpStatus :: Int

The response's http status code.

newImportApiResponse

Arguments

:: Int
$sel:httpStatus:ImportApiResponse'

-> ImportApiResponse

Create a value of ImportApiResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:ImportApiResponse', importApiResponse_apiId - The API ID.

$sel:disableExecuteApiEndpoint:ImportApiResponse', importApiResponse_disableExecuteApiEndpoint - Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

$sel:apiEndpoint:ImportApiResponse', importApiResponse_apiEndpoint - The URI of the API, of the form {api-id}.execute-api.{region}.amazonaws.com. The stage name is typically appended to this URI to form a complete path to a deployed API stage.

$sel:warnings:ImportApiResponse', importApiResponse_warnings - The warning messages reported when failonwarnings is turned on during API import.

$sel:createdDate:ImportApiResponse', importApiResponse_createdDate - The timestamp when the API was created.

$sel:name:ImportApiResponse', importApiResponse_name - The name of the API.

$sel:version:ImportApiResponse', importApiResponse_version - A version identifier for the API.

$sel:apiGatewayManaged:ImportApiResponse', importApiResponse_apiGatewayManaged - Specifies whether an API is managed by API Gateway. You can't update or delete a managed API by using API Gateway. A managed API can be deleted only through the tooling or service that created it.

$sel:apiKeySelectionExpression:ImportApiResponse', importApiResponse_apiKeySelectionExpression - An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

$sel:corsConfiguration:ImportApiResponse', importApiResponse_corsConfiguration - A CORS configuration. Supported only for HTTP APIs.

$sel:routeSelectionExpression:ImportApiResponse', importApiResponse_routeSelectionExpression - The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

$sel:importInfo:ImportApiResponse', importApiResponse_importInfo - The validation information during API import. This may include particular properties of your OpenAPI definition which are ignored during import. Supported only for HTTP APIs.

$sel:disableSchemaValidation:ImportApiResponse', importApiResponse_disableSchemaValidation - Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

$sel:description:ImportApiResponse', importApiResponse_description - The description of the API.

$sel:protocolType:ImportApiResponse', importApiResponse_protocolType - The API protocol.

$sel:tags:ImportApiResponse', importApiResponse_tags - A collection of tags associated with the API.

$sel:httpStatus:ImportApiResponse', importApiResponse_httpStatus - The response's http status code.

Response Lenses

importApiResponse_apiId :: Lens' ImportApiResponse (Maybe Text)

The API ID.

importApiResponse_disableExecuteApiEndpoint :: Lens' ImportApiResponse (Maybe Bool)

Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

importApiResponse_apiEndpoint :: Lens' ImportApiResponse (Maybe Text)

The URI of the API, of the form {api-id}.execute-api.{region}.amazonaws.com. The stage name is typically appended to this URI to form a complete path to a deployed API stage.

importApiResponse_warnings :: Lens' ImportApiResponse (Maybe [Text])

The warning messages reported when failonwarnings is turned on during API import.

importApiResponse_createdDate :: Lens' ImportApiResponse (Maybe UTCTime)

The timestamp when the API was created.

importApiResponse_name :: Lens' ImportApiResponse (Maybe Text)

The name of the API.

importApiResponse_version :: Lens' ImportApiResponse (Maybe Text)

A version identifier for the API.

importApiResponse_apiGatewayManaged :: Lens' ImportApiResponse (Maybe Bool)

Specifies whether an API is managed by API Gateway. You can't update or delete a managed API by using API Gateway. A managed API can be deleted only through the tooling or service that created it.

importApiResponse_apiKeySelectionExpression :: Lens' ImportApiResponse (Maybe Text)

An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

importApiResponse_corsConfiguration :: Lens' ImportApiResponse (Maybe Cors)

A CORS configuration. Supported only for HTTP APIs.

importApiResponse_routeSelectionExpression :: Lens' ImportApiResponse (Maybe Text)

The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be ${request.method} ${request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

importApiResponse_importInfo :: Lens' ImportApiResponse (Maybe [Text])

The validation information during API import. This may include particular properties of your OpenAPI definition which are ignored during import. Supported only for HTTP APIs.

importApiResponse_disableSchemaValidation :: Lens' ImportApiResponse (Maybe Bool)

Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

importApiResponse_description :: Lens' ImportApiResponse (Maybe Text)

The description of the API.

importApiResponse_protocolType :: Lens' ImportApiResponse (Maybe ProtocolType)

The API protocol.

importApiResponse_tags :: Lens' ImportApiResponse (Maybe (HashMap Text Text))

A collection of tags associated with the API.

importApiResponse_httpStatus :: Lens' ImportApiResponse Int

The response's http status code.