Network.AWS.ApiGatewayV2.Types.RouteResponse

Description

Documentation

data RouteResponse

Represents a route response.

See: newRouteResponse smart constructor.

Constructors

RouteResponse'

Fields

modelSelectionExpression :: Maybe Text

Represents the model selection expression of a route response. Supported only for WebSocket APIs.

responseModels :: Maybe (HashMap Text Text)

Represents the response models of a route response.

routeResponseId :: Maybe Text

Represents the identifier of a route response.

responseParameters :: Maybe (HashMap Text ParameterConstraints)

Represents the response parameters of a route response.

routeResponseKey :: Text

Represents the route response key of a route response.

newRouteResponse

Arguments

:: Text
$sel:routeResponseKey:RouteResponse'

-> RouteResponse

Create a value of RouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelSelectionExpression:RouteResponse', routeResponse_modelSelectionExpression - Represents the model selection expression of a route response. Supported only for WebSocket APIs.

$sel:responseModels:RouteResponse', routeResponse_responseModels - Represents the response models of a route response.

$sel:routeResponseId:RouteResponse', routeResponse_routeResponseId - Represents the identifier of a route response.

$sel:responseParameters:RouteResponse', routeResponse_responseParameters - Represents the response parameters of a route response.

$sel:routeResponseKey:RouteResponse', routeResponse_routeResponseKey - Represents the route response key of a route response.

routeResponse_modelSelectionExpression :: Lens' RouteResponse (Maybe Text)

Represents the model selection expression of a route response. Supported only for WebSocket APIs.

routeResponse_responseModels :: Lens' RouteResponse (Maybe (HashMap Text Text))

Represents the response models of a route response.

routeResponse_routeResponseId :: Lens' RouteResponse (Maybe Text)

Represents the identifier of a route response.

routeResponse_responseParameters :: Lens' RouteResponse (Maybe (HashMap Text ParameterConstraints))

Represents the response parameters of a route response.

routeResponse_routeResponseKey :: Lens' RouteResponse Text

Represents the route response key of a route response.