Network.AWS.ApiGatewayV2.Types.Stage

Description

Documentation

data Stage

Represents an API stage.

See: newStage smart constructor.

Constructors

Stage'

Fields

lastDeploymentStatusMessage :: Maybe Text

Describes the status of the last deployment of a stage. Supported only for stages with autoDeploy enabled.

deploymentId :: Maybe Text

The identifier of the Deployment that the Stage is associated with. Can't be updated if autoDeploy is enabled.

routeSettings :: Maybe (HashMap Text RouteSettings)

Route settings for the stage, by routeKey.

accessLogSettings :: Maybe AccessLogSettings

Settings for logging access in this stage.

clientCertificateId :: Maybe Text

The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

stageVariables :: Maybe (HashMap Text Text)

A map that defines the stage variables for a stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

autoDeploy :: Maybe Bool

Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

createdDate :: Maybe POSIX

The timestamp when the stage was created.

defaultRouteSettings :: Maybe RouteSettings

Default route settings for the stage.

apiGatewayManaged :: Maybe Bool

Specifies whether a stage is managed by API Gateway. If you created an API using quick create, the $default stage is managed by API Gateway. You can't modify the $default stage.

lastUpdatedDate :: Maybe POSIX

The timestamp when the stage was last updated.

description :: Maybe Text

The description of the stage.

tags :: Maybe (HashMap Text Text)

The collection of tags. Each tag element is associated with a given resource.

stageName :: Text

The name of the stage.

newStage

Arguments

:: Text
$sel:stageName:Stage'

-> Stage

Create a value of Stage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastDeploymentStatusMessage:Stage', stage_lastDeploymentStatusMessage - Describes the status of the last deployment of a stage. Supported only for stages with autoDeploy enabled.

$sel:deploymentId:Stage', stage_deploymentId - The identifier of the Deployment that the Stage is associated with. Can't be updated if autoDeploy is enabled.

$sel:routeSettings:Stage', stage_routeSettings - Route settings for the stage, by routeKey.

$sel:accessLogSettings:Stage', stage_accessLogSettings - Settings for logging access in this stage.

$sel:clientCertificateId:Stage', stage_clientCertificateId - The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

$sel:stageVariables:Stage', stage_stageVariables - A map that defines the stage variables for a stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

$sel:autoDeploy:Stage', stage_autoDeploy - Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

$sel:createdDate:Stage', stage_createdDate - The timestamp when the stage was created.

$sel:defaultRouteSettings:Stage', stage_defaultRouteSettings - Default route settings for the stage.

$sel:apiGatewayManaged:Stage', stage_apiGatewayManaged - Specifies whether a stage is managed by API Gateway. If you created an API using quick create, the $default stage is managed by API Gateway. You can't modify the $default stage.

$sel:lastUpdatedDate:Stage', stage_lastUpdatedDate - The timestamp when the stage was last updated.

$sel:description:Stage', stage_description - The description of the stage.

$sel:tags:Stage', stage_tags - The collection of tags. Each tag element is associated with a given resource.

$sel:stageName:Stage', stage_stageName - The name of the stage.

stage_lastDeploymentStatusMessage :: Lens' Stage (Maybe Text)

Describes the status of the last deployment of a stage. Supported only for stages with autoDeploy enabled.

stage_deploymentId :: Lens' Stage (Maybe Text)

The identifier of the Deployment that the Stage is associated with. Can't be updated if autoDeploy is enabled.

stage_routeSettings :: Lens' Stage (Maybe (HashMap Text RouteSettings))

Route settings for the stage, by routeKey.

stage_accessLogSettings :: Lens' Stage (Maybe AccessLogSettings)

Settings for logging access in this stage.

stage_clientCertificateId :: Lens' Stage (Maybe Text)

The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

stage_stageVariables :: Lens' Stage (Maybe (HashMap Text Text))

A map that defines the stage variables for a stage resource. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&=,]+.

stage_autoDeploy :: Lens' Stage (Maybe Bool)

Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

stage_createdDate :: Lens' Stage (Maybe UTCTime)

The timestamp when the stage was created.

stage_defaultRouteSettings :: Lens' Stage (Maybe RouteSettings)

Default route settings for the stage.

stage_apiGatewayManaged :: Lens' Stage (Maybe Bool)

Specifies whether a stage is managed by API Gateway. If you created an API using quick create, the $default stage is managed by API Gateway. You can't modify the $default stage.

stage_lastUpdatedDate :: Lens' Stage (Maybe UTCTime)

The timestamp when the stage was last updated.

stage_description :: Lens' Stage (Maybe Text)

The description of the stage.

stage_tags :: Lens' Stage (Maybe (HashMap Text Text))

The collection of tags. Each tag element is associated with a given resource.

stage_stageName :: Lens' Stage Text

The name of the stage.