Network.AWS.ApiGatewayV2.Types.VpcLink

Description

Documentation

data VpcLink

Represents a VPC link.

See: newVpcLink smart constructor.

Constructors

VpcLink'

Fields

createdDate :: Maybe POSIX

The timestamp when the VPC link was created.

vpcLinkVersion :: Maybe VpcLinkVersion

The version of the VPC link.

vpcLinkStatusMessage :: Maybe Text

A message summarizing the cause of the status of the VPC link.

tags :: Maybe (HashMap Text Text)

Tags for the VPC link.

vpcLinkStatus :: Maybe VpcLinkStatus

The status of the VPC link.

vpcLinkId :: Text

The ID of the VPC link.

securityGroupIds :: [Text]

A list of security group IDs for the VPC link.

subnetIds :: [Text]

A list of subnet IDs to include in the VPC link.

name :: Text

The name of the VPC link.

newVpcLink

Arguments

:: Text
$sel:vpcLinkId:VpcLink'

-> Text
$sel:name:VpcLink'

-> VpcLink

Create a value of VpcLink with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:VpcLink', vpcLink_createdDate - The timestamp when the VPC link was created.

$sel:vpcLinkVersion:VpcLink', vpcLink_vpcLinkVersion - The version of the VPC link.

$sel:vpcLinkStatusMessage:VpcLink', vpcLink_vpcLinkStatusMessage - A message summarizing the cause of the status of the VPC link.

$sel:tags:VpcLink', vpcLink_tags - Tags for the VPC link.

$sel:vpcLinkStatus:VpcLink', vpcLink_vpcLinkStatus - The status of the VPC link.

$sel:vpcLinkId:VpcLink', vpcLink_vpcLinkId - The ID of the VPC link.

$sel:securityGroupIds:VpcLink', vpcLink_securityGroupIds - A list of security group IDs for the VPC link.

$sel:subnetIds:VpcLink', vpcLink_subnetIds - A list of subnet IDs to include in the VPC link.

$sel:name:VpcLink', vpcLink_name - The name of the VPC link.

vpcLink_createdDate :: Lens' VpcLink (Maybe UTCTime)

The timestamp when the VPC link was created.

vpcLink_vpcLinkVersion :: Lens' VpcLink (Maybe VpcLinkVersion)

The version of the VPC link.

vpcLink_vpcLinkStatusMessage :: Lens' VpcLink (Maybe Text)

A message summarizing the cause of the status of the VPC link.

vpcLink_tags :: Lens' VpcLink (Maybe (HashMap Text Text))

Tags for the VPC link.

vpcLink_vpcLinkStatus :: Lens' VpcLink (Maybe VpcLinkStatus)

The status of the VPC link.

vpcLink_vpcLinkId :: Lens' VpcLink Text

The ID of the VPC link.

vpcLink_securityGroupIds :: Lens' VpcLink [Text]

A list of security group IDs for the VPC link.

vpcLink_subnetIds :: Lens' VpcLink [Text]

A list of subnet IDs to include in the VPC link.

vpcLink_name :: Lens' VpcLink Text

The name of the VPC link.