Network.AWS.ApiGatewayV2.UpdateAuthorizer

Description

Updates an Authorizer.

Creating a Request

data UpdateAuthorizer

Updates an Authorizer.

See: newUpdateAuthorizer smart constructor.

Constructors

UpdateAuthorizer'

Fields

authorizerCredentialsArn :: Maybe Text

Specifies the required credentials as an IAM role for API Gateway to invoke the authorizer. To specify an IAM role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To use resource-based permissions on the Lambda function, don't specify this parameter.

identityValidationExpression :: Maybe Text

This parameter is not used.

enableSimpleResponses :: Maybe Bool

Specifies whether a Lambda authorizer returns a response in a simple format. By default, a Lambda authorizer must return an IAM policy. If enabled, the Lambda authorizer can return a boolean value instead of an IAM policy. Supported only for HTTP APIs. To learn more, see Working with AWS Lambda authorizers for HTTP APIs

authorizerUri :: Maybe Text

The authorizer's Uniform Resource Identifier (URI). For REQUEST authorizers, this must be a well-formed Lambda function URI, for example, arn:aws:apigateway:us-west-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-west-2:{account_id}:function:{lambda_function_name}/invocations. In general, the URI has this form: arn:aws:apigateway:{region}:lambda:path/{service_api} , where {region} is the same as the region hosting the Lambda function, path indicates that the remaining substring in the URI should be treated as the path to the resource, including the initial /. For Lambda functions, this is usually of the form /2015-03-31/functions/[FunctionARN]/invocations. Supported only for REQUEST authorizers.

authorizerPayloadFormatVersion :: Maybe Text

Specifies the format of the payload sent to an HTTP API Lambda authorizer. Required for HTTP API Lambda authorizers. Supported values are 1.0 and 2.0. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

jwtConfiguration :: Maybe JWTConfiguration

Represents the configuration of a JWT authorizer. Required for the JWT authorizer type. Supported only for HTTP APIs.

name :: Maybe Text

The name of the authorizer.

authorizerResultTtlInSeconds :: Maybe Natural

The time to live (TTL) for cached authorizer results, in seconds. If it equals 0, authorization caching is disabled. If it is greater than 0, API Gateway caches authorizer responses. The maximum value is 3600, or 1 hour. Supported only for HTTP API Lambda authorizers.

identitySource :: Maybe [Text]

The identity source for which authorization is requested.

For a REQUEST authorizer, this is optional. The value is a set of one or more mapping expressions of the specified request parameters. The identity source can be headers, query string parameters, stage variables, and context parameters. For example, if an Auth header and a Name query string parameter are defined as identity sources, this value is route.request.header.Auth, route.request.querystring.Name for WebSocket APIs. For HTTP APIs, use selection expressions prefixed with $, for example, $request.header.Auth, $request.querystring.Name. These parameters are used to perform runtime validation for Lambda-based authorizers by verifying all of the identity-related request parameters are present in the request, not null, and non-empty. Only when this is true does the authorizer invoke the authorizer Lambda function. Otherwise, it returns a 401 Unauthorized response without calling the Lambda function. For HTTP APIs, identity sources are also used as the cache key when caching is enabled. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

For JWT, a single entry that specifies where to extract the JSON Web Token (JWT) from inbound requests. Currently only header-based and query parameter-based selections are supported, for example $request.header.Authorization.

authorizerType :: Maybe AuthorizerType

The authorizer type. Specify REQUEST for a Lambda function using incoming request parameters. Specify JWT to use JSON Web Tokens (supported only for HTTP APIs).

authorizerId :: Text

The authorizer identifier.

apiId :: Text

The API identifier.

newUpdateAuthorizer

Arguments

:: Text
$sel:authorizerId:UpdateAuthorizer'

-> Text
$sel:apiId:UpdateAuthorizer'

-> UpdateAuthorizer

Create a value of UpdateAuthorizer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizerCredentialsArn:UpdateAuthorizer', updateAuthorizer_authorizerCredentialsArn - Specifies the required credentials as an IAM role for API Gateway to invoke the authorizer. To specify an IAM role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To use resource-based permissions on the Lambda function, don't specify this parameter.

$sel:identityValidationExpression:UpdateAuthorizer', updateAuthorizer_identityValidationExpression - This parameter is not used.

$sel:enableSimpleResponses:UpdateAuthorizer', updateAuthorizer_enableSimpleResponses - Specifies whether a Lambda authorizer returns a response in a simple format. By default, a Lambda authorizer must return an IAM policy. If enabled, the Lambda authorizer can return a boolean value instead of an IAM policy. Supported only for HTTP APIs. To learn more, see Working with AWS Lambda authorizers for HTTP APIs

$sel:authorizerUri:UpdateAuthorizer', updateAuthorizer_authorizerUri - The authorizer's Uniform Resource Identifier (URI). For REQUEST authorizers, this must be a well-formed Lambda function URI, for example, arn:aws:apigateway:us-west-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-west-2:{account_id}:function:{lambda_function_name}/invocations. In general, the URI has this form: arn:aws:apigateway:{region}:lambda:path/{service_api} , where {region} is the same as the region hosting the Lambda function, path indicates that the remaining substring in the URI should be treated as the path to the resource, including the initial /. For Lambda functions, this is usually of the form /2015-03-31/functions/[FunctionARN]/invocations. Supported only for REQUEST authorizers.

$sel:authorizerPayloadFormatVersion:UpdateAuthorizer', updateAuthorizer_authorizerPayloadFormatVersion - Specifies the format of the payload sent to an HTTP API Lambda authorizer. Required for HTTP API Lambda authorizers. Supported values are 1.0 and 2.0. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

$sel:jwtConfiguration:UpdateAuthorizer', updateAuthorizer_jwtConfiguration - Represents the configuration of a JWT authorizer. Required for the JWT authorizer type. Supported only for HTTP APIs.

$sel:name:UpdateAuthorizer', updateAuthorizer_name - The name of the authorizer.

$sel:authorizerResultTtlInSeconds:UpdateAuthorizer', updateAuthorizer_authorizerResultTtlInSeconds - The time to live (TTL) for cached authorizer results, in seconds. If it equals 0, authorization caching is disabled. If it is greater than 0, API Gateway caches authorizer responses. The maximum value is 3600, or 1 hour. Supported only for HTTP API Lambda authorizers.

$sel:identitySource:UpdateAuthorizer', updateAuthorizer_identitySource - The identity source for which authorization is requested.

For a REQUEST authorizer, this is optional. The value is a set of one or more mapping expressions of the specified request parameters. The identity source can be headers, query string parameters, stage variables, and context parameters. For example, if an Auth header and a Name query string parameter are defined as identity sources, this value is route.request.header.Auth, route.request.querystring.Name for WebSocket APIs. For HTTP APIs, use selection expressions prefixed with $, for example, $request.header.Auth, $request.querystring.Name. These parameters are used to perform runtime validation for Lambda-based authorizers by verifying all of the identity-related request parameters are present in the request, not null, and non-empty. Only when this is true does the authorizer invoke the authorizer Lambda function. Otherwise, it returns a 401 Unauthorized response without calling the Lambda function. For HTTP APIs, identity sources are also used as the cache key when caching is enabled. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

For JWT, a single entry that specifies where to extract the JSON Web Token (JWT) from inbound requests. Currently only header-based and query parameter-based selections are supported, for example $request.header.Authorization.

$sel:authorizerType:UpdateAuthorizer', updateAuthorizer_authorizerType - The authorizer type. Specify REQUEST for a Lambda function using incoming request parameters. Specify JWT to use JSON Web Tokens (supported only for HTTP APIs).

$sel:authorizerId:UpdateAuthorizer', updateAuthorizer_authorizerId - The authorizer identifier.

$sel:apiId:UpdateAuthorizer', updateAuthorizer_apiId - The API identifier.

Request Lenses

updateAuthorizer_authorizerCredentialsArn :: Lens' UpdateAuthorizer (Maybe Text)

Specifies the required credentials as an IAM role for API Gateway to invoke the authorizer. To specify an IAM role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To use resource-based permissions on the Lambda function, don't specify this parameter.

updateAuthorizer_identityValidationExpression :: Lens' UpdateAuthorizer (Maybe Text)

This parameter is not used.

updateAuthorizer_enableSimpleResponses :: Lens' UpdateAuthorizer (Maybe Bool)

Specifies whether a Lambda authorizer returns a response in a simple format. By default, a Lambda authorizer must return an IAM policy. If enabled, the Lambda authorizer can return a boolean value instead of an IAM policy. Supported only for HTTP APIs. To learn more, see Working with AWS Lambda authorizers for HTTP APIs

updateAuthorizer_authorizerUri :: Lens' UpdateAuthorizer (Maybe Text)

The authorizer's Uniform Resource Identifier (URI). For REQUEST authorizers, this must be a well-formed Lambda function URI, for example, arn:aws:apigateway:us-west-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-west-2:{account_id}:function:{lambda_function_name}/invocations. In general, the URI has this form: arn:aws:apigateway:{region}:lambda:path/{service_api} , where {region} is the same as the region hosting the Lambda function, path indicates that the remaining substring in the URI should be treated as the path to the resource, including the initial /. For Lambda functions, this is usually of the form /2015-03-31/functions/[FunctionARN]/invocations. Supported only for REQUEST authorizers.

updateAuthorizer_authorizerPayloadFormatVersion :: Lens' UpdateAuthorizer (Maybe Text)

Specifies the format of the payload sent to an HTTP API Lambda authorizer. Required for HTTP API Lambda authorizers. Supported values are 1.0 and 2.0. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

updateAuthorizer_jwtConfiguration :: Lens' UpdateAuthorizer (Maybe JWTConfiguration)

Represents the configuration of a JWT authorizer. Required for the JWT authorizer type. Supported only for HTTP APIs.

updateAuthorizer_name :: Lens' UpdateAuthorizer (Maybe Text)

The name of the authorizer.

updateAuthorizer_authorizerResultTtlInSeconds :: Lens' UpdateAuthorizer (Maybe Natural)

The time to live (TTL) for cached authorizer results, in seconds. If it equals 0, authorization caching is disabled. If it is greater than 0, API Gateway caches authorizer responses. The maximum value is 3600, or 1 hour. Supported only for HTTP API Lambda authorizers.

updateAuthorizer_identitySource :: Lens' UpdateAuthorizer (Maybe [Text])

The identity source for which authorization is requested.

For a REQUEST authorizer, this is optional. The value is a set of one or more mapping expressions of the specified request parameters. The identity source can be headers, query string parameters, stage variables, and context parameters. For example, if an Auth header and a Name query string parameter are defined as identity sources, this value is route.request.header.Auth, route.request.querystring.Name for WebSocket APIs. For HTTP APIs, use selection expressions prefixed with $, for example, $request.header.Auth, $request.querystring.Name. These parameters are used to perform runtime validation for Lambda-based authorizers by verifying all of the identity-related request parameters are present in the request, not null, and non-empty. Only when this is true does the authorizer invoke the authorizer Lambda function. Otherwise, it returns a 401 Unauthorized response without calling the Lambda function. For HTTP APIs, identity sources are also used as the cache key when caching is enabled. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

For JWT, a single entry that specifies where to extract the JSON Web Token (JWT) from inbound requests. Currently only header-based and query parameter-based selections are supported, for example $request.header.Authorization.

updateAuthorizer_authorizerType :: Lens' UpdateAuthorizer (Maybe AuthorizerType)

The authorizer type. Specify REQUEST for a Lambda function using incoming request parameters. Specify JWT to use JSON Web Tokens (supported only for HTTP APIs).

updateAuthorizer_authorizerId :: Lens' UpdateAuthorizer Text

The authorizer identifier.

updateAuthorizer_apiId :: Lens' UpdateAuthorizer Text

The API identifier.

Destructuring the Response

data UpdateAuthorizerResponse

See: newUpdateAuthorizerResponse smart constructor.

Constructors

UpdateAuthorizerResponse'

Fields

authorizerCredentialsArn :: Maybe Text

Specifies the required credentials as an IAM role for API Gateway to invoke the authorizer. To specify an IAM role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To use resource-based permissions on the Lambda function, don't specify this parameter. Supported only for REQUEST authorizers.

identityValidationExpression :: Maybe Text

The validation expression does not apply to the REQUEST authorizer.

enableSimpleResponses :: Maybe Bool

Specifies whether a Lambda authorizer returns a response in a simple format. If enabled, the Lambda authorizer can return a boolean value instead of an IAM policy. Supported only for HTTP APIs. To learn more, see Working with AWS Lambda authorizers for HTTP APIs

authorizerUri :: Maybe Text

The authorizer's Uniform Resource Identifier (URI). For REQUEST authorizers, this must be a well-formed Lambda function URI, for example, arn:aws:apigateway:us-west-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-west-2:{account_id}:function:{lambda_function_name}/invocations. In general, the URI has this form: arn:aws:apigateway:{region}:lambda:path/{service_api} , where {region} is the same as the region hosting the Lambda function, path indicates that the remaining substring in the URI should be treated as the path to the resource, including the initial /. For Lambda functions, this is usually of the form /2015-03-31/functions/[FunctionARN]/invocations. Supported only for REQUEST authorizers.

authorizerPayloadFormatVersion :: Maybe Text

Specifies the format of the payload sent to an HTTP API Lambda authorizer. Required for HTTP API Lambda authorizers. Supported values are 1.0 and 2.0. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

jwtConfiguration :: Maybe JWTConfiguration

Represents the configuration of a JWT authorizer. Required for the JWT authorizer type. Supported only for HTTP APIs.

authorizerId :: Maybe Text

The authorizer identifier.

name :: Maybe Text

The name of the authorizer.

authorizerResultTtlInSeconds :: Maybe Natural

The time to live (TTL) for cached authorizer results, in seconds. If it equals 0, authorization caching is disabled. If it is greater than 0, API Gateway caches authorizer responses. The maximum value is 3600, or 1 hour. Supported only for HTTP API Lambda authorizers.

identitySource :: Maybe [Text]

The identity source for which authorization is requested.

For a REQUEST authorizer, this is optional. The value is a set of one or more mapping expressions of the specified request parameters. The identity source can be headers, query string parameters, stage variables, and context parameters. For example, if an Auth header and a Name query string parameter are defined as identity sources, this value is route.request.header.Auth, route.request.querystring.Name for WebSocket APIs. For HTTP APIs, use selection expressions prefixed with $, for example, $request.header.Auth, $request.querystring.Name. These parameters are used to perform runtime validation for Lambda-based authorizers by verifying all of the identity-related request parameters are present in the request, not null, and non-empty. Only when this is true does the authorizer invoke the authorizer Lambda function. Otherwise, it returns a 401 Unauthorized response without calling the Lambda function. For HTTP APIs, identity sources are also used as the cache key when caching is enabled. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

For JWT, a single entry that specifies where to extract the JSON Web Token (JWT) from inbound requests. Currently only header-based and query parameter-based selections are supported, for example $request.header.Authorization.

authorizerType :: Maybe AuthorizerType

The authorizer type. Specify REQUEST for a Lambda function using incoming request parameters. Specify JWT to use JSON Web Tokens (supported only for HTTP APIs).

httpStatus :: Int

The response's http status code.

newUpdateAuthorizerResponse

Arguments

:: Int
$sel:httpStatus:UpdateAuthorizerResponse'

-> UpdateAuthorizerResponse

Create a value of UpdateAuthorizerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizerCredentialsArn:UpdateAuthorizer', updateAuthorizerResponse_authorizerCredentialsArn - Specifies the required credentials as an IAM role for API Gateway to invoke the authorizer. To specify an IAM role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To use resource-based permissions on the Lambda function, don't specify this parameter. Supported only for REQUEST authorizers.

$sel:identityValidationExpression:UpdateAuthorizer', updateAuthorizerResponse_identityValidationExpression - The validation expression does not apply to the REQUEST authorizer.

$sel:enableSimpleResponses:UpdateAuthorizer', updateAuthorizerResponse_enableSimpleResponses - Specifies whether a Lambda authorizer returns a response in a simple format. If enabled, the Lambda authorizer can return a boolean value instead of an IAM policy. Supported only for HTTP APIs. To learn more, see Working with AWS Lambda authorizers for HTTP APIs

$sel:authorizerUri:UpdateAuthorizer', updateAuthorizerResponse_authorizerUri - The authorizer's Uniform Resource Identifier (URI). For REQUEST authorizers, this must be a well-formed Lambda function URI, for example, arn:aws:apigateway:us-west-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-west-2:{account_id}:function:{lambda_function_name}/invocations. In general, the URI has this form: arn:aws:apigateway:{region}:lambda:path/{service_api} , where {region} is the same as the region hosting the Lambda function, path indicates that the remaining substring in the URI should be treated as the path to the resource, including the initial /. For Lambda functions, this is usually of the form /2015-03-31/functions/[FunctionARN]/invocations. Supported only for REQUEST authorizers.

$sel:authorizerPayloadFormatVersion:UpdateAuthorizer', updateAuthorizerResponse_authorizerPayloadFormatVersion - Specifies the format of the payload sent to an HTTP API Lambda authorizer. Required for HTTP API Lambda authorizers. Supported values are 1.0 and 2.0. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

$sel:jwtConfiguration:UpdateAuthorizer', updateAuthorizerResponse_jwtConfiguration - Represents the configuration of a JWT authorizer. Required for the JWT authorizer type. Supported only for HTTP APIs.

$sel:authorizerId:UpdateAuthorizer', updateAuthorizerResponse_authorizerId - The authorizer identifier.

$sel:name:UpdateAuthorizer', updateAuthorizerResponse_name - The name of the authorizer.

$sel:authorizerResultTtlInSeconds:UpdateAuthorizer', updateAuthorizerResponse_authorizerResultTtlInSeconds - The time to live (TTL) for cached authorizer results, in seconds. If it equals 0, authorization caching is disabled. If it is greater than 0, API Gateway caches authorizer responses. The maximum value is 3600, or 1 hour. Supported only for HTTP API Lambda authorizers.

$sel:identitySource:UpdateAuthorizer', updateAuthorizerResponse_identitySource - The identity source for which authorization is requested.

For a REQUEST authorizer, this is optional. The value is a set of one or more mapping expressions of the specified request parameters. The identity source can be headers, query string parameters, stage variables, and context parameters. For example, if an Auth header and a Name query string parameter are defined as identity sources, this value is route.request.header.Auth, route.request.querystring.Name for WebSocket APIs. For HTTP APIs, use selection expressions prefixed with $, for example, $request.header.Auth, $request.querystring.Name. These parameters are used to perform runtime validation for Lambda-based authorizers by verifying all of the identity-related request parameters are present in the request, not null, and non-empty. Only when this is true does the authorizer invoke the authorizer Lambda function. Otherwise, it returns a 401 Unauthorized response without calling the Lambda function. For HTTP APIs, identity sources are also used as the cache key when caching is enabled. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

For JWT, a single entry that specifies where to extract the JSON Web Token (JWT) from inbound requests. Currently only header-based and query parameter-based selections are supported, for example $request.header.Authorization.

$sel:authorizerType:UpdateAuthorizer', updateAuthorizerResponse_authorizerType - The authorizer type. Specify REQUEST for a Lambda function using incoming request parameters. Specify JWT to use JSON Web Tokens (supported only for HTTP APIs).

$sel:httpStatus:UpdateAuthorizerResponse', updateAuthorizerResponse_httpStatus - The response's http status code.

Response Lenses

updateAuthorizerResponse_authorizerCredentialsArn :: Lens' UpdateAuthorizerResponse (Maybe Text)

Specifies the required credentials as an IAM role for API Gateway to invoke the authorizer. To specify an IAM role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To use resource-based permissions on the Lambda function, don't specify this parameter. Supported only for REQUEST authorizers.

updateAuthorizerResponse_identityValidationExpression :: Lens' UpdateAuthorizerResponse (Maybe Text)

The validation expression does not apply to the REQUEST authorizer.

updateAuthorizerResponse_enableSimpleResponses :: Lens' UpdateAuthorizerResponse (Maybe Bool)

Specifies whether a Lambda authorizer returns a response in a simple format. If enabled, the Lambda authorizer can return a boolean value instead of an IAM policy. Supported only for HTTP APIs. To learn more, see Working with AWS Lambda authorizers for HTTP APIs

updateAuthorizerResponse_authorizerUri :: Lens' UpdateAuthorizerResponse (Maybe Text)

The authorizer's Uniform Resource Identifier (URI). For REQUEST authorizers, this must be a well-formed Lambda function URI, for example, arn:aws:apigateway:us-west-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-west-2:{account_id}:function:{lambda_function_name}/invocations. In general, the URI has this form: arn:aws:apigateway:{region}:lambda:path/{service_api} , where {region} is the same as the region hosting the Lambda function, path indicates that the remaining substring in the URI should be treated as the path to the resource, including the initial /. For Lambda functions, this is usually of the form /2015-03-31/functions/[FunctionARN]/invocations. Supported only for REQUEST authorizers.

updateAuthorizerResponse_authorizerPayloadFormatVersion :: Lens' UpdateAuthorizerResponse (Maybe Text)

Specifies the format of the payload sent to an HTTP API Lambda authorizer. Required for HTTP API Lambda authorizers. Supported values are 1.0 and 2.0. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

updateAuthorizerResponse_jwtConfiguration :: Lens' UpdateAuthorizerResponse (Maybe JWTConfiguration)

Represents the configuration of a JWT authorizer. Required for the JWT authorizer type. Supported only for HTTP APIs.

updateAuthorizerResponse_authorizerId :: Lens' UpdateAuthorizerResponse (Maybe Text)

The authorizer identifier.

updateAuthorizerResponse_name :: Lens' UpdateAuthorizerResponse (Maybe Text)

The name of the authorizer.

updateAuthorizerResponse_authorizerResultTtlInSeconds :: Lens' UpdateAuthorizerResponse (Maybe Natural)

The time to live (TTL) for cached authorizer results, in seconds. If it equals 0, authorization caching is disabled. If it is greater than 0, API Gateway caches authorizer responses. The maximum value is 3600, or 1 hour. Supported only for HTTP API Lambda authorizers.

updateAuthorizerResponse_identitySource :: Lens' UpdateAuthorizerResponse (Maybe [Text])

The identity source for which authorization is requested.

For a REQUEST authorizer, this is optional. The value is a set of one or more mapping expressions of the specified request parameters. The identity source can be headers, query string parameters, stage variables, and context parameters. For example, if an Auth header and a Name query string parameter are defined as identity sources, this value is route.request.header.Auth, route.request.querystring.Name for WebSocket APIs. For HTTP APIs, use selection expressions prefixed with $, for example, $request.header.Auth, $request.querystring.Name. These parameters are used to perform runtime validation for Lambda-based authorizers by verifying all of the identity-related request parameters are present in the request, not null, and non-empty. Only when this is true does the authorizer invoke the authorizer Lambda function. Otherwise, it returns a 401 Unauthorized response without calling the Lambda function. For HTTP APIs, identity sources are also used as the cache key when caching is enabled. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

For JWT, a single entry that specifies where to extract the JSON Web Token (JWT) from inbound requests. Currently only header-based and query parameter-based selections are supported, for example $request.header.Authorization.

updateAuthorizerResponse_authorizerType :: Lens' UpdateAuthorizerResponse (Maybe AuthorizerType)

The authorizer type. Specify REQUEST for a Lambda function using incoming request parameters. Specify JWT to use JSON Web Tokens (supported only for HTTP APIs).

updateAuthorizerResponse_httpStatus :: Lens' UpdateAuthorizerResponse Int

The response's http status code.