Network.AWS.ApiGatewayV2.UpdateDomainName

Description

Updates a domain name.

Creating a Request

data UpdateDomainName

Updates a DomainName.

See: newUpdateDomainName smart constructor.

Constructors

UpdateDomainName'

Fields

domainNameConfigurations :: Maybe [DomainNameConfiguration]

The domain name configurations.

mutualTlsAuthentication :: Maybe MutualTlsAuthenticationInput

The mutual TLS authentication configuration for a custom domain name.

domainName :: Text

The domain name.

newUpdateDomainName

Arguments

:: Text
$sel:domainName:UpdateDomainName'

-> UpdateDomainName

Create a value of UpdateDomainName with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainNameConfigurations:UpdateDomainName', updateDomainName_domainNameConfigurations - The domain name configurations.

$sel:mutualTlsAuthentication:UpdateDomainName', updateDomainName_mutualTlsAuthentication - The mutual TLS authentication configuration for a custom domain name.

$sel:domainName:UpdateDomainName', updateDomainName_domainName - The domain name.

Request Lenses

updateDomainName_domainNameConfigurations :: Lens' UpdateDomainName (Maybe [DomainNameConfiguration])

The domain name configurations.

updateDomainName_mutualTlsAuthentication :: Lens' UpdateDomainName (Maybe MutualTlsAuthenticationInput)

The mutual TLS authentication configuration for a custom domain name.

updateDomainName_domainName :: Lens' UpdateDomainName Text

The domain name.

Destructuring the Response

data UpdateDomainNameResponse

See: newUpdateDomainNameResponse smart constructor.

Constructors

UpdateDomainNameResponse'

Fields

domainNameConfigurations :: Maybe [DomainNameConfiguration]

The domain name configurations.

domainName :: Maybe Text

The name of the DomainName resource.

mutualTlsAuthentication :: Maybe MutualTlsAuthentication

The mutual TLS authentication configuration for a custom domain name.

apiMappingSelectionExpression :: Maybe Text

The API mapping selection expression.

tags :: Maybe (HashMap Text Text)

The collection of tags associated with a domain name.

httpStatus :: Int

The response's http status code.

newUpdateDomainNameResponse

Arguments

:: Int
$sel:httpStatus:UpdateDomainNameResponse'

-> UpdateDomainNameResponse

Create a value of UpdateDomainNameResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainNameConfigurations:UpdateDomainName', updateDomainNameResponse_domainNameConfigurations - The domain name configurations.

$sel:domainName:UpdateDomainName', updateDomainNameResponse_domainName - The name of the DomainName resource.

$sel:mutualTlsAuthentication:UpdateDomainName', updateDomainNameResponse_mutualTlsAuthentication - The mutual TLS authentication configuration for a custom domain name.

$sel:apiMappingSelectionExpression:UpdateDomainNameResponse', updateDomainNameResponse_apiMappingSelectionExpression - The API mapping selection expression.

$sel:tags:UpdateDomainNameResponse', updateDomainNameResponse_tags - The collection of tags associated with a domain name.

$sel:httpStatus:UpdateDomainNameResponse', updateDomainNameResponse_httpStatus - The response's http status code.

Response Lenses

updateDomainNameResponse_domainNameConfigurations :: Lens' UpdateDomainNameResponse (Maybe [DomainNameConfiguration])

The domain name configurations.

updateDomainNameResponse_domainName :: Lens' UpdateDomainNameResponse (Maybe Text)

The name of the DomainName resource.

updateDomainNameResponse_mutualTlsAuthentication :: Lens' UpdateDomainNameResponse (Maybe MutualTlsAuthentication)

The mutual TLS authentication configuration for a custom domain name.

updateDomainNameResponse_apiMappingSelectionExpression :: Lens' UpdateDomainNameResponse (Maybe Text)

The API mapping selection expression.

updateDomainNameResponse_tags :: Lens' UpdateDomainNameResponse (Maybe (HashMap Text Text))

The collection of tags associated with a domain name.

updateDomainNameResponse_httpStatus :: Lens' UpdateDomainNameResponse Int

The response's http status code.