Network.AWS.ApiGatewayV2.UpdateIntegrationResponse

Description

Updates an IntegrationResponses.

Creating a Request

data UpdateIntegrationResponse

Updates an IntegrationResponses.

See: newUpdateIntegrationResponse smart constructor.

Constructors

UpdateIntegrationResponse'

Fields

integrationResponseKey :: Maybe Text

The integration response key.

templateSelectionExpression :: Maybe Text

The template selection expression for the integration response. Supported only for WebSocket APIs.

contentHandlingStrategy :: Maybe ContentHandlingStrategy

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

responseTemplates :: Maybe (HashMap Text Text)

The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

responseParameters :: Maybe (HashMap Text Text)

A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name} , where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression} , where {name} is a valid and unique response header name and {JSON-expression} is a valid JSON expression without the $ prefix.

apiId :: Text

The API identifier.

integrationResponseId :: Text

The integration response ID.

integrationId :: Text

The integration ID.

newUpdateIntegrationResponse

Arguments

:: Text
$sel:apiId:UpdateIntegrationResponse'

-> Text
$sel:integrationResponseId:UpdateIntegrationResponse'

-> Text
$sel:integrationId:UpdateIntegrationResponse'

-> UpdateIntegrationResponse

Create a value of UpdateIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:integrationResponseKey:UpdateIntegrationResponse', updateIntegrationResponse_integrationResponseKey - The integration response key.

$sel:templateSelectionExpression:UpdateIntegrationResponse', updateIntegrationResponse_templateSelectionExpression - The template selection expression for the integration response. Supported only for WebSocket APIs.

$sel:contentHandlingStrategy:UpdateIntegrationResponse', updateIntegrationResponse_contentHandlingStrategy - Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

$sel:responseTemplates:UpdateIntegrationResponse', updateIntegrationResponse_responseTemplates - The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

$sel:responseParameters:UpdateIntegrationResponse', updateIntegrationResponse_responseParameters - A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name} , where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression} , where {name} is a valid and unique response header name and {JSON-expression} is a valid JSON expression without the $ prefix.

$sel:apiId:UpdateIntegrationResponse', updateIntegrationResponse_apiId - The API identifier.

$sel:integrationResponseId:UpdateIntegrationResponse', updateIntegrationResponse_integrationResponseId - The integration response ID.

$sel:integrationId:UpdateIntegrationResponse', updateIntegrationResponse_integrationId - The integration ID.

Request Lenses

updateIntegrationResponse_integrationResponseKey :: Lens' UpdateIntegrationResponse (Maybe Text)

The integration response key.

updateIntegrationResponse_templateSelectionExpression :: Lens' UpdateIntegrationResponse (Maybe Text)

The template selection expression for the integration response. Supported only for WebSocket APIs.

updateIntegrationResponse_contentHandlingStrategy :: Lens' UpdateIntegrationResponse (Maybe ContentHandlingStrategy)

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

updateIntegrationResponse_responseTemplates :: Lens' UpdateIntegrationResponse (Maybe (HashMap Text Text))

The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

updateIntegrationResponse_responseParameters :: Lens' UpdateIntegrationResponse (Maybe (HashMap Text Text))

A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name} , where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression} , where {name} is a valid and unique response header name and {JSON-expression} is a valid JSON expression without the $ prefix.

updateIntegrationResponse_apiId :: Lens' UpdateIntegrationResponse Text

The API identifier.

updateIntegrationResponse_integrationResponseId :: Lens' UpdateIntegrationResponse Text

The integration response ID.

updateIntegrationResponse_integrationId :: Lens' UpdateIntegrationResponse Text

The integration ID.

Destructuring the Response

data UpdateIntegrationResponseResponse

See: newUpdateIntegrationResponseResponse smart constructor.

Constructors

UpdateIntegrationResponseResponse'

Fields

integrationResponseId :: Maybe Text

The integration response ID.

integrationResponseKey :: Maybe Text

The integration response key.

templateSelectionExpression :: Maybe Text

The template selection expressions for the integration response.

contentHandlingStrategy :: Maybe ContentHandlingStrategy

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

responseTemplates :: Maybe (HashMap Text Text)

The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

responseParameters :: Maybe (HashMap Text Text)

A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

httpStatus :: Int

The response's http status code.

newUpdateIntegrationResponseResponse

Arguments

:: Int
$sel:httpStatus:UpdateIntegrationResponseResponse'

-> UpdateIntegrationResponseResponse

Create a value of UpdateIntegrationResponseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:integrationResponseId:UpdateIntegrationResponse', updateIntegrationResponseResponse_integrationResponseId - The integration response ID.

$sel:integrationResponseKey:UpdateIntegrationResponse', updateIntegrationResponseResponse_integrationResponseKey - The integration response key.

$sel:templateSelectionExpression:UpdateIntegrationResponse', updateIntegrationResponseResponse_templateSelectionExpression - The template selection expressions for the integration response.

$sel:contentHandlingStrategy:UpdateIntegrationResponse', updateIntegrationResponseResponse_contentHandlingStrategy - Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

$sel:responseTemplates:UpdateIntegrationResponse', updateIntegrationResponseResponse_responseTemplates - The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

$sel:responseParameters:UpdateIntegrationResponse', updateIntegrationResponseResponse_responseParameters - A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

$sel:httpStatus:UpdateIntegrationResponseResponse', updateIntegrationResponseResponse_httpStatus - The response's http status code.

Response Lenses

updateIntegrationResponseResponse_integrationResponseId :: Lens' UpdateIntegrationResponseResponse (Maybe Text)

The integration response ID.

updateIntegrationResponseResponse_integrationResponseKey :: Lens' UpdateIntegrationResponseResponse (Maybe Text)

The integration response key.

updateIntegrationResponseResponse_templateSelectionExpression :: Lens' UpdateIntegrationResponseResponse (Maybe Text)

The template selection expressions for the integration response.

updateIntegrationResponseResponse_contentHandlingStrategy :: Lens' UpdateIntegrationResponseResponse (Maybe ContentHandlingStrategy)

Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERT_TO_BINARY and CONVERT_TO_TEXT, with the following behaviors:

CONVERT_TO_BINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob.

CONVERT_TO_TEXT: Converts a response payload from a binary blob to a Base64-encoded string.

If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

updateIntegrationResponseResponse_responseTemplates :: Lens' UpdateIntegrationResponseResponse (Maybe (HashMap Text Text))

The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

updateIntegrationResponseResponse_responseParameters :: Lens' UpdateIntegrationResponseResponse (Maybe (HashMap Text Text))

A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where name is a valid and unique response header name and JSON-expression is a valid JSON expression without the $ prefix.

updateIntegrationResponseResponse_httpStatus :: Lens' UpdateIntegrationResponseResponse Int

The response's http status code.