Network.AWS.ApiGatewayV2.UpdateRouteResponse

Description

Updates a RouteResponse.

Creating a Request

data UpdateRouteResponse

Updates a RouteResponse.

See: newUpdateRouteResponse smart constructor.

Constructors

UpdateRouteResponse'

Fields

modelSelectionExpression :: Maybe Text

The model selection expression for the route response. Supported only for WebSocket APIs.

responseModels :: Maybe (HashMap Text Text)

The response models for the route response.

routeResponseKey :: Maybe Text

The route response key.

responseParameters :: Maybe (HashMap Text ParameterConstraints)

The route response parameters.

routeResponseId :: Text

The route response ID.

apiId :: Text

The API identifier.

routeId :: Text

The route ID.

newUpdateRouteResponse

Arguments

:: Text
$sel:routeResponseId:UpdateRouteResponse'

-> Text
$sel:apiId:UpdateRouteResponse'

-> Text
$sel:routeId:UpdateRouteResponse'

-> UpdateRouteResponse

Create a value of UpdateRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelSelectionExpression:UpdateRouteResponse', updateRouteResponse_modelSelectionExpression - The model selection expression for the route response. Supported only for WebSocket APIs.

$sel:responseModels:UpdateRouteResponse', updateRouteResponse_responseModels - The response models for the route response.

$sel:routeResponseKey:UpdateRouteResponse', updateRouteResponse_routeResponseKey - The route response key.

$sel:responseParameters:UpdateRouteResponse', updateRouteResponse_responseParameters - The route response parameters.

$sel:routeResponseId:UpdateRouteResponse', updateRouteResponse_routeResponseId - The route response ID.

$sel:apiId:UpdateRouteResponse', updateRouteResponse_apiId - The API identifier.

$sel:routeId:UpdateRouteResponse', updateRouteResponse_routeId - The route ID.

Request Lenses

updateRouteResponse_modelSelectionExpression :: Lens' UpdateRouteResponse (Maybe Text)

The model selection expression for the route response. Supported only for WebSocket APIs.

updateRouteResponse_responseModels :: Lens' UpdateRouteResponse (Maybe (HashMap Text Text))

The response models for the route response.

updateRouteResponse_routeResponseKey :: Lens' UpdateRouteResponse (Maybe Text)

The route response key.

updateRouteResponse_responseParameters :: Lens' UpdateRouteResponse (Maybe (HashMap Text ParameterConstraints))

The route response parameters.

updateRouteResponse_routeResponseId :: Lens' UpdateRouteResponse Text

The route response ID.

updateRouteResponse_apiId :: Lens' UpdateRouteResponse Text

The API identifier.

updateRouteResponse_routeId :: Lens' UpdateRouteResponse Text

The route ID.

Destructuring the Response

data UpdateRouteResponseResponse

See: newUpdateRouteResponseResponse smart constructor.

Constructors

UpdateRouteResponseResponse'

Fields

modelSelectionExpression :: Maybe Text

Represents the model selection expression of a route response. Supported only for WebSocket APIs.

responseModels :: Maybe (HashMap Text Text)

Represents the response models of a route response.

routeResponseId :: Maybe Text

Represents the identifier of a route response.

routeResponseKey :: Maybe Text

Represents the route response key of a route response.

responseParameters :: Maybe (HashMap Text ParameterConstraints)

Represents the response parameters of a route response.

httpStatus :: Int

The response's http status code.

newUpdateRouteResponseResponse

Arguments

:: Int
$sel:httpStatus:UpdateRouteResponseResponse'

-> UpdateRouteResponseResponse

Create a value of UpdateRouteResponseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelSelectionExpression:UpdateRouteResponse', updateRouteResponseResponse_modelSelectionExpression - Represents the model selection expression of a route response. Supported only for WebSocket APIs.

$sel:responseModels:UpdateRouteResponse', updateRouteResponseResponse_responseModels - Represents the response models of a route response.

$sel:routeResponseId:UpdateRouteResponse', updateRouteResponseResponse_routeResponseId - Represents the identifier of a route response.

$sel:routeResponseKey:UpdateRouteResponse', updateRouteResponseResponse_routeResponseKey - Represents the route response key of a route response.

$sel:responseParameters:UpdateRouteResponse', updateRouteResponseResponse_responseParameters - Represents the response parameters of a route response.

$sel:httpStatus:UpdateRouteResponseResponse', updateRouteResponseResponse_httpStatus - The response's http status code.

Response Lenses

updateRouteResponseResponse_modelSelectionExpression :: Lens' UpdateRouteResponseResponse (Maybe Text)

Represents the model selection expression of a route response. Supported only for WebSocket APIs.

updateRouteResponseResponse_responseModels :: Lens' UpdateRouteResponseResponse (Maybe (HashMap Text Text))

Represents the response models of a route response.

updateRouteResponseResponse_routeResponseId :: Lens' UpdateRouteResponseResponse (Maybe Text)

Represents the identifier of a route response.

updateRouteResponseResponse_routeResponseKey :: Lens' UpdateRouteResponseResponse (Maybe Text)

Represents the route response key of a route response.

updateRouteResponseResponse_responseParameters :: Lens' UpdateRouteResponseResponse (Maybe (HashMap Text ParameterConstraints))

Represents the response parameters of a route response.

updateRouteResponseResponse_httpStatus :: Lens' UpdateRouteResponseResponse Int

The response's http status code.