Network.AWS.AppConfig.CreateDeploymentStrategy

Description

A deployment strategy defines important criteria for rolling out your configuration to the designated targets. A deployment strategy includes: the overall duration required, a percentage of targets to receive the deployment during each interval, an algorithm that defines how percentage grows, and bake time.

Creating a Request

data CreateDeploymentStrategy

See: newCreateDeploymentStrategy smart constructor.

Constructors

CreateDeploymentStrategy'

Fields

finalBakeTimeInMinutes :: Maybe Natural

The amount of time AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

description :: Maybe Text

A description of the deployment strategy.

growthType :: Maybe GrowthType

The algorithm used to define how percentage grows over time. AWS AppConfig supports the following growth types:

Linear: For this type, AppConfig processes the deployment by dividing the total number of targets by the value specified for Step percentage. For example, a linear deployment that uses a Step percentage of 10 deploys the configuration to 10 percent of the hosts. After those deployments are complete, the system deploys the configuration to the next 10 percent. This continues until 100% of the targets have successfully received the configuration.

Exponential: For this type, AppConfig processes the deployment exponentially using the following formula: G*(2^N). In this formula, G is the growth factor specified by the user and N is the number of steps until the configuration is deployed to all targets. For example, if you specify a growth factor of 2, then the system rolls out the configuration as follows:

2*(2^0)

2*(2^1)

2*(2^2)

Expressed numerically, the deployment rolls out as follows: 2% of the targets, 4% of the targets, 8% of the targets, and continues until the configuration has been deployed to all targets.

tags :: Maybe (HashMap Text Text)

Metadata to assign to the deployment strategy. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

name :: Text

A name for the deployment strategy.

deploymentDurationInMinutes :: Natural

Total amount of time for a deployment to last.

growthFactor :: Double

The percentage of targets to receive a deployed configuration during each interval.

replicateTo :: ReplicateTo

Save the deployment strategy to a Systems Manager (SSM) document.

newCreateDeploymentStrategy

Arguments

:: Text
$sel:name:CreateDeploymentStrategy'

-> Natural
$sel:deploymentDurationInMinutes:CreateDeploymentStrategy'

-> Double
$sel:growthFactor:CreateDeploymentStrategy'

-> ReplicateTo
$sel:replicateTo:CreateDeploymentStrategy'

-> CreateDeploymentStrategy

Create a value of CreateDeploymentStrategy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:finalBakeTimeInMinutes:CreateDeploymentStrategy', createDeploymentStrategy_finalBakeTimeInMinutes - The amount of time AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

$sel:description:CreateDeploymentStrategy', createDeploymentStrategy_description - A description of the deployment strategy.

$sel:growthType:CreateDeploymentStrategy', createDeploymentStrategy_growthType - The algorithm used to define how percentage grows over time. AWS AppConfig supports the following growth types:

Linear: For this type, AppConfig processes the deployment by dividing the total number of targets by the value specified for Step percentage. For example, a linear deployment that uses a Step percentage of 10 deploys the configuration to 10 percent of the hosts. After those deployments are complete, the system deploys the configuration to the next 10 percent. This continues until 100% of the targets have successfully received the configuration.

Exponential: For this type, AppConfig processes the deployment exponentially using the following formula: G*(2^N). In this formula, G is the growth factor specified by the user and N is the number of steps until the configuration is deployed to all targets. For example, if you specify a growth factor of 2, then the system rolls out the configuration as follows:

2*(2^0)

2*(2^1)

2*(2^2)

Expressed numerically, the deployment rolls out as follows: 2% of the targets, 4% of the targets, 8% of the targets, and continues until the configuration has been deployed to all targets.

$sel:tags:CreateDeploymentStrategy', createDeploymentStrategy_tags - Metadata to assign to the deployment strategy. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

$sel:name:CreateDeploymentStrategy', createDeploymentStrategy_name - A name for the deployment strategy.

$sel:deploymentDurationInMinutes:CreateDeploymentStrategy', createDeploymentStrategy_deploymentDurationInMinutes - Total amount of time for a deployment to last.

$sel:growthFactor:CreateDeploymentStrategy', createDeploymentStrategy_growthFactor - The percentage of targets to receive a deployed configuration during each interval.

$sel:replicateTo:CreateDeploymentStrategy', createDeploymentStrategy_replicateTo - Save the deployment strategy to a Systems Manager (SSM) document.

Request Lenses

createDeploymentStrategy_finalBakeTimeInMinutes :: Lens' CreateDeploymentStrategy (Maybe Natural)

The amount of time AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

createDeploymentStrategy_description :: Lens' CreateDeploymentStrategy (Maybe Text)

A description of the deployment strategy.

createDeploymentStrategy_growthType :: Lens' CreateDeploymentStrategy (Maybe GrowthType)

The algorithm used to define how percentage grows over time. AWS AppConfig supports the following growth types:

Linear: For this type, AppConfig processes the deployment by dividing the total number of targets by the value specified for Step percentage. For example, a linear deployment that uses a Step percentage of 10 deploys the configuration to 10 percent of the hosts. After those deployments are complete, the system deploys the configuration to the next 10 percent. This continues until 100% of the targets have successfully received the configuration.

Exponential: For this type, AppConfig processes the deployment exponentially using the following formula: G*(2^N). In this formula, G is the growth factor specified by the user and N is the number of steps until the configuration is deployed to all targets. For example, if you specify a growth factor of 2, then the system rolls out the configuration as follows:

2*(2^0)

2*(2^1)

2*(2^2)

Expressed numerically, the deployment rolls out as follows: 2% of the targets, 4% of the targets, 8% of the targets, and continues until the configuration has been deployed to all targets.

createDeploymentStrategy_tags :: Lens' CreateDeploymentStrategy (Maybe (HashMap Text Text))

Metadata to assign to the deployment strategy. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

createDeploymentStrategy_name :: Lens' CreateDeploymentStrategy Text

A name for the deployment strategy.

createDeploymentStrategy_deploymentDurationInMinutes :: Lens' CreateDeploymentStrategy Natural

Total amount of time for a deployment to last.

createDeploymentStrategy_growthFactor :: Lens' CreateDeploymentStrategy Double

The percentage of targets to receive a deployed configuration during each interval.

createDeploymentStrategy_replicateTo :: Lens' CreateDeploymentStrategy ReplicateTo

Save the deployment strategy to a Systems Manager (SSM) document.

Destructuring the Response

data DeploymentStrategy

See: newDeploymentStrategy smart constructor.

Constructors

DeploymentStrategy'

Fields

growthFactor :: Maybe Double

The percentage of targets that received a deployed configuration during each interval.

replicateTo :: Maybe ReplicateTo

Save the deployment strategy to a Systems Manager (SSM) document.

name :: Maybe Text

The name of the deployment strategy.

id :: Maybe Text

The deployment strategy ID.

deploymentDurationInMinutes :: Maybe Natural

Total amount of time the deployment lasted.

finalBakeTimeInMinutes :: Maybe Natural

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

description :: Maybe Text

The description of the deployment strategy.

growthType :: Maybe GrowthType

The algorithm used to define how percentage grew over time.

newDeploymentStrategy :: DeploymentStrategy

Create a value of DeploymentStrategy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:growthFactor:DeploymentStrategy', deploymentStrategy_growthFactor - The percentage of targets that received a deployed configuration during each interval.

$sel:replicateTo:DeploymentStrategy', deploymentStrategy_replicateTo - Save the deployment strategy to a Systems Manager (SSM) document.

$sel:name:DeploymentStrategy', deploymentStrategy_name - The name of the deployment strategy.

$sel:id:DeploymentStrategy', deploymentStrategy_id - The deployment strategy ID.

$sel:deploymentDurationInMinutes:DeploymentStrategy', deploymentStrategy_deploymentDurationInMinutes - Total amount of time the deployment lasted.

$sel:finalBakeTimeInMinutes:DeploymentStrategy', deploymentStrategy_finalBakeTimeInMinutes - The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

$sel:description:DeploymentStrategy', deploymentStrategy_description - The description of the deployment strategy.

$sel:growthType:DeploymentStrategy', deploymentStrategy_growthType - The algorithm used to define how percentage grew over time.

Response Lenses

deploymentStrategy_growthFactor :: Lens' DeploymentStrategy (Maybe Double)

The percentage of targets that received a deployed configuration during each interval.

deploymentStrategy_replicateTo :: Lens' DeploymentStrategy (Maybe ReplicateTo)

Save the deployment strategy to a Systems Manager (SSM) document.

deploymentStrategy_name :: Lens' DeploymentStrategy (Maybe Text)

The name of the deployment strategy.

deploymentStrategy_id :: Lens' DeploymentStrategy (Maybe Text)

The deployment strategy ID.

deploymentStrategy_deploymentDurationInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

Total amount of time the deployment lasted.

deploymentStrategy_finalBakeTimeInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deploymentStrategy_description :: Lens' DeploymentStrategy (Maybe Text)

The description of the deployment strategy.

deploymentStrategy_growthType :: Lens' DeploymentStrategy (Maybe GrowthType)

The algorithm used to define how percentage grew over time.