Network.AWS.AppConfig.CreateEnvironment

Description

For each application, you define one or more environments. An environment is a logical deployment group of AppConfig targets, such as applications in a Beta or Production environment. You can also define environments for application subcomponents such as the Web, Mobile and Back-end components for your application. You can configure Amazon CloudWatch alarms for each environment. The system monitors alarms during a configuration deployment. If an alarm is triggered, the system rolls back the configuration.

Creating a Request

data CreateEnvironment

See: newCreateEnvironment smart constructor.

Constructors

CreateEnvironment'

Fields

monitors :: Maybe [Monitor]

Amazon CloudWatch alarms to monitor during the deployment process.

description :: Maybe Text

A description of the environment.

tags :: Maybe (HashMap Text Text)

Metadata to assign to the environment. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

applicationId :: Text

The application ID.

name :: Text

A name for the environment.

newCreateEnvironment

Arguments

:: Text
$sel:applicationId:CreateEnvironment'

-> Text
$sel:name:CreateEnvironment'

-> CreateEnvironment

Create a value of CreateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monitors:CreateEnvironment', createEnvironment_monitors - Amazon CloudWatch alarms to monitor during the deployment process.

$sel:description:CreateEnvironment', createEnvironment_description - A description of the environment.

$sel:tags:CreateEnvironment', createEnvironment_tags - Metadata to assign to the environment. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

$sel:applicationId:CreateEnvironment', createEnvironment_applicationId - The application ID.

$sel:name:CreateEnvironment', createEnvironment_name - A name for the environment.

Request Lenses

createEnvironment_monitors :: Lens' CreateEnvironment (Maybe [Monitor])

Amazon CloudWatch alarms to monitor during the deployment process.

createEnvironment_description :: Lens' CreateEnvironment (Maybe Text)

A description of the environment.

createEnvironment_tags :: Lens' CreateEnvironment (Maybe (HashMap Text Text))

Metadata to assign to the environment. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

createEnvironment_applicationId :: Lens' CreateEnvironment Text

The application ID.

createEnvironment_name :: Lens' CreateEnvironment Text

A name for the environment.

Destructuring the Response

data Environment

See: newEnvironment smart constructor.

Constructors

Environment'

Fields

state :: Maybe EnvironmentState

The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

monitors :: Maybe [Monitor]

Amazon CloudWatch alarms monitored during the deployment.

applicationId :: Maybe Text

The application ID.

name :: Maybe Text

The name of the environment.

id :: Maybe Text

The environment ID.

description :: Maybe Text

The description of the environment.

newEnvironment :: Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Environment', environment_state - The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

$sel:monitors:Environment', environment_monitors - Amazon CloudWatch alarms monitored during the deployment.

$sel:applicationId:Environment', environment_applicationId - The application ID.

$sel:name:Environment', environment_name - The name of the environment.

$sel:id:Environment', environment_id - The environment ID.

$sel:description:Environment', environment_description - The description of the environment.

Response Lenses

environment_state :: Lens' Environment (Maybe EnvironmentState)

The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

environment_monitors :: Lens' Environment (Maybe [Monitor])

Amazon CloudWatch alarms monitored during the deployment.

environment_applicationId :: Lens' Environment (Maybe Text)

The application ID.

environment_name :: Lens' Environment (Maybe Text)

The name of the environment.

environment_id :: Lens' Environment (Maybe Text)

The environment ID.

environment_description :: Lens' Environment (Maybe Text)

The description of the environment.