Network.AWS.AppConfig.DeleteDeploymentStrategy

Description

Delete a deployment strategy. Deleting a deployment strategy does not delete a configuration from a host.

Creating a Request

data DeleteDeploymentStrategy

See: newDeleteDeploymentStrategy smart constructor.

Constructors

DeleteDeploymentStrategy'

Fields

deploymentStrategyId :: Text

The ID of the deployment strategy you want to delete.

newDeleteDeploymentStrategy

Arguments

:: Text
$sel:deploymentStrategyId:DeleteDeploymentStrategy'

-> DeleteDeploymentStrategy

Create a value of DeleteDeploymentStrategy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentStrategyId:DeleteDeploymentStrategy', deleteDeploymentStrategy_deploymentStrategyId - The ID of the deployment strategy you want to delete.

Request Lenses

deleteDeploymentStrategy_deploymentStrategyId :: Lens' DeleteDeploymentStrategy Text

The ID of the deployment strategy you want to delete.

Destructuring the Response

data DeleteDeploymentStrategyResponse

See: newDeleteDeploymentStrategyResponse smart constructor.

Constructors

DeleteDeploymentStrategyResponse'

newDeleteDeploymentStrategyResponse :: DeleteDeploymentStrategyResponse

Create a value of DeleteDeploymentStrategyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.