Network.AWS.AppConfig.GetConfiguration

Description

Receive information about a configuration.

AWS AppConfig uses the value of the ClientConfigurationVersion parameter to identify the configuration version on your clients. If you don’t send ClientConfigurationVersion with each call to GetConfiguration, your clients receive the current configuration. You are charged each time your clients receive a configuration.

To avoid excess charges, we recommend that you include the ClientConfigurationVersion value with every call to GetConfiguration. This value must be saved on your client. Subsequent calls to GetConfiguration must pass this value by using the ClientConfigurationVersion parameter.

Creating a Request

data GetConfiguration

See: newGetConfiguration smart constructor.

Constructors

GetConfiguration'

Fields

clientConfigurationVersion :: Maybe Text

The configuration version returned in the most recent GetConfiguration response.

AWS AppConfig uses the value of the ClientConfigurationVersion parameter to identify the configuration version on your clients. If you don’t send ClientConfigurationVersion with each call to GetConfiguration, your clients receive the current configuration. You are charged each time your clients receive a configuration.

To avoid excess charges, we recommend that you include the ClientConfigurationVersion value with every call to GetConfiguration. This value must be saved on your client. Subsequent calls to GetConfiguration must pass this value by using the ClientConfigurationVersion parameter.

For more information about working with configurations, see Retrieving the Configuration in the AWS AppConfig User Guide.

application :: Text

The application to get. Specify either the application name or the application ID.

environment :: Text

The environment to get. Specify either the environment name or the environment ID.

configuration :: Text

The configuration to get. Specify either the configuration name or the configuration ID.

clientId :: Text

A unique ID to identify the client for the configuration. This ID enables AppConfig to deploy the configuration in intervals, as defined in the deployment strategy.

newGetConfiguration

Arguments

:: Text
$sel:application:GetConfiguration'

-> Text
$sel:environment:GetConfiguration'

-> Text
$sel:configuration:GetConfiguration'

-> Text
$sel:clientId:GetConfiguration'

-> GetConfiguration

Create a value of GetConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientConfigurationVersion:GetConfiguration', getConfiguration_clientConfigurationVersion - The configuration version returned in the most recent GetConfiguration response.

AWS AppConfig uses the value of the ClientConfigurationVersion parameter to identify the configuration version on your clients. If you don’t send ClientConfigurationVersion with each call to GetConfiguration, your clients receive the current configuration. You are charged each time your clients receive a configuration.

To avoid excess charges, we recommend that you include the ClientConfigurationVersion value with every call to GetConfiguration. This value must be saved on your client. Subsequent calls to GetConfiguration must pass this value by using the ClientConfigurationVersion parameter.

For more information about working with configurations, see Retrieving the Configuration in the AWS AppConfig User Guide.

$sel:application:GetConfiguration', getConfiguration_application - The application to get. Specify either the application name or the application ID.

$sel:environment:GetConfiguration', getConfiguration_environment - The environment to get. Specify either the environment name or the environment ID.

$sel:configuration:GetConfiguration', getConfiguration_configuration - The configuration to get. Specify either the configuration name or the configuration ID.

$sel:clientId:GetConfiguration', getConfiguration_clientId - A unique ID to identify the client for the configuration. This ID enables AppConfig to deploy the configuration in intervals, as defined in the deployment strategy.

Request Lenses

getConfiguration_clientConfigurationVersion :: Lens' GetConfiguration (Maybe Text)

The configuration version returned in the most recent GetConfiguration response.

AWS AppConfig uses the value of the ClientConfigurationVersion parameter to identify the configuration version on your clients. If you don’t send ClientConfigurationVersion with each call to GetConfiguration, your clients receive the current configuration. You are charged each time your clients receive a configuration.

To avoid excess charges, we recommend that you include the ClientConfigurationVersion value with every call to GetConfiguration. This value must be saved on your client. Subsequent calls to GetConfiguration must pass this value by using the ClientConfigurationVersion parameter.

For more information about working with configurations, see Retrieving the Configuration in the AWS AppConfig User Guide.

getConfiguration_application :: Lens' GetConfiguration Text

The application to get. Specify either the application name or the application ID.

getConfiguration_environment :: Lens' GetConfiguration Text

The environment to get. Specify either the environment name or the environment ID.

getConfiguration_configuration :: Lens' GetConfiguration Text

The configuration to get. Specify either the configuration name or the configuration ID.

getConfiguration_clientId :: Lens' GetConfiguration Text

A unique ID to identify the client for the configuration. This ID enables AppConfig to deploy the configuration in intervals, as defined in the deployment strategy.

Destructuring the Response

data GetConfigurationResponse

See: newGetConfigurationResponse smart constructor.

Constructors

GetConfigurationResponse'

Fields

configurationVersion :: Maybe Text

The configuration version.

content :: Maybe (Sensitive ByteString)

The content of the configuration or the configuration data.

contentType :: Maybe Text

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

httpStatus :: Int

The response's http status code.

newGetConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetConfigurationResponse'

-> GetConfigurationResponse

Create a value of GetConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationVersion:GetConfigurationResponse', getConfigurationResponse_configurationVersion - The configuration version.

$sel:content:GetConfigurationResponse', getConfigurationResponse_content - The content of the configuration or the configuration data.

$sel:contentType:GetConfigurationResponse', getConfigurationResponse_contentType - A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

$sel:httpStatus:GetConfigurationResponse', getConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getConfigurationResponse_configurationVersion :: Lens' GetConfigurationResponse (Maybe Text)

The configuration version.

getConfigurationResponse_content :: Lens' GetConfigurationResponse (Maybe ByteString)

The content of the configuration or the configuration data.

getConfigurationResponse_contentType :: Lens' GetConfigurationResponse (Maybe Text)

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

getConfigurationResponse_httpStatus :: Lens' GetConfigurationResponse Int

The response's http status code.