Network.AWS.AppConfig.GetDeploymentStrategy

Description

Retrieve information about a deployment strategy. A deployment strategy defines important criteria for rolling out your configuration to the designated targets. A deployment strategy includes: the overall duration required, a percentage of targets to receive the deployment during each interval, an algorithm that defines how percentage grows, and bake time.

Creating a Request

data GetDeploymentStrategy

See: newGetDeploymentStrategy smart constructor.

Constructors

GetDeploymentStrategy'

Fields

deploymentStrategyId :: Text

The ID of the deployment strategy to get.

newGetDeploymentStrategy

Arguments

:: Text
$sel:deploymentStrategyId:GetDeploymentStrategy'

-> GetDeploymentStrategy

Create a value of GetDeploymentStrategy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentStrategyId:GetDeploymentStrategy', getDeploymentStrategy_deploymentStrategyId - The ID of the deployment strategy to get.

Request Lenses

getDeploymentStrategy_deploymentStrategyId :: Lens' GetDeploymentStrategy Text

The ID of the deployment strategy to get.

Destructuring the Response

data DeploymentStrategy

See: newDeploymentStrategy smart constructor.

Constructors

DeploymentStrategy'

Fields

growthFactor :: Maybe Double

The percentage of targets that received a deployed configuration during each interval.

replicateTo :: Maybe ReplicateTo

Save the deployment strategy to a Systems Manager (SSM) document.

name :: Maybe Text

The name of the deployment strategy.

id :: Maybe Text

The deployment strategy ID.

deploymentDurationInMinutes :: Maybe Natural

Total amount of time the deployment lasted.

finalBakeTimeInMinutes :: Maybe Natural

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

description :: Maybe Text

The description of the deployment strategy.

growthType :: Maybe GrowthType

The algorithm used to define how percentage grew over time.

newDeploymentStrategy :: DeploymentStrategy

Create a value of DeploymentStrategy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:growthFactor:DeploymentStrategy', deploymentStrategy_growthFactor - The percentage of targets that received a deployed configuration during each interval.

$sel:replicateTo:DeploymentStrategy', deploymentStrategy_replicateTo - Save the deployment strategy to a Systems Manager (SSM) document.

$sel:name:DeploymentStrategy', deploymentStrategy_name - The name of the deployment strategy.

$sel:id:DeploymentStrategy', deploymentStrategy_id - The deployment strategy ID.

$sel:deploymentDurationInMinutes:DeploymentStrategy', deploymentStrategy_deploymentDurationInMinutes - Total amount of time the deployment lasted.

$sel:finalBakeTimeInMinutes:DeploymentStrategy', deploymentStrategy_finalBakeTimeInMinutes - The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

$sel:description:DeploymentStrategy', deploymentStrategy_description - The description of the deployment strategy.

$sel:growthType:DeploymentStrategy', deploymentStrategy_growthType - The algorithm used to define how percentage grew over time.

Response Lenses

deploymentStrategy_growthFactor :: Lens' DeploymentStrategy (Maybe Double)

The percentage of targets that received a deployed configuration during each interval.

deploymentStrategy_replicateTo :: Lens' DeploymentStrategy (Maybe ReplicateTo)

Save the deployment strategy to a Systems Manager (SSM) document.

deploymentStrategy_name :: Lens' DeploymentStrategy (Maybe Text)

The name of the deployment strategy.

deploymentStrategy_id :: Lens' DeploymentStrategy (Maybe Text)

The deployment strategy ID.

deploymentStrategy_deploymentDurationInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

Total amount of time the deployment lasted.

deploymentStrategy_finalBakeTimeInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deploymentStrategy_description :: Lens' DeploymentStrategy (Maybe Text)

The description of the deployment strategy.

deploymentStrategy_growthType :: Lens' DeploymentStrategy (Maybe GrowthType)

The algorithm used to define how percentage grew over time.