Network.AWS.AppConfig.GetEnvironment

Description

Retrieve information about an environment. An environment is a logical deployment group of AppConfig applications, such as applications in a Production environment or in an EU_Region environment. Each configuration deployment targets an environment. You can enable one or more Amazon CloudWatch alarms for an environment. If an alarm is triggered during a deployment, AppConfig roles back the configuration.

Creating a Request

data GetEnvironment

See: newGetEnvironment smart constructor.

Constructors

GetEnvironment'

Fields

applicationId :: Text

The ID of the application that includes the environment you want to get.

environmentId :: Text

The ID of the environment you wnat to get.

newGetEnvironment

Arguments

:: Text
$sel:applicationId:GetEnvironment'

-> Text
$sel:environmentId:GetEnvironment'

-> GetEnvironment

Create a value of GetEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:GetEnvironment', getEnvironment_applicationId - The ID of the application that includes the environment you want to get.

$sel:environmentId:GetEnvironment', getEnvironment_environmentId - The ID of the environment you wnat to get.

Request Lenses

getEnvironment_applicationId :: Lens' GetEnvironment Text

The ID of the application that includes the environment you want to get.

getEnvironment_environmentId :: Lens' GetEnvironment Text

The ID of the environment you wnat to get.

Destructuring the Response

data Environment

See: newEnvironment smart constructor.

Constructors

Environment'

Fields

state :: Maybe EnvironmentState

The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

monitors :: Maybe [Monitor]

Amazon CloudWatch alarms monitored during the deployment.

applicationId :: Maybe Text

The application ID.

name :: Maybe Text

The name of the environment.

id :: Maybe Text

The environment ID.

description :: Maybe Text

The description of the environment.

newEnvironment :: Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Environment', environment_state - The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

$sel:monitors:Environment', environment_monitors - Amazon CloudWatch alarms monitored during the deployment.

$sel:applicationId:Environment', environment_applicationId - The application ID.

$sel:name:Environment', environment_name - The name of the environment.

$sel:id:Environment', environment_id - The environment ID.

$sel:description:Environment', environment_description - The description of the environment.

Response Lenses

environment_state :: Lens' Environment (Maybe EnvironmentState)

The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

environment_monitors :: Lens' Environment (Maybe [Monitor])

Amazon CloudWatch alarms monitored during the deployment.

environment_applicationId :: Lens' Environment (Maybe Text)

The application ID.

environment_name :: Lens' Environment (Maybe Text)

The name of the environment.

environment_id :: Lens' Environment (Maybe Text)

The environment ID.

environment_description :: Lens' Environment (Maybe Text)

The description of the environment.