Network.AWS.AppConfig.ListDeployments

Description

Lists the deployments for an environment.

Creating a Request

data ListDeployments

See: newListDeployments smart constructor.

Constructors

ListDeployments'

Fields

nextToken :: Maybe Text

A token to start the list. Use this token to get the next set of results.

maxResults :: Maybe Natural

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

applicationId :: Text

The application ID.

environmentId :: Text

The environment ID.

newListDeployments

Arguments

:: Text
$sel:applicationId:ListDeployments'

-> Text
$sel:environmentId:ListDeployments'

-> ListDeployments

Create a value of ListDeployments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDeployments', listDeployments_nextToken - A token to start the list. Use this token to get the next set of results.

$sel:maxResults:ListDeployments', listDeployments_maxResults - The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

$sel:applicationId:ListDeployments', listDeployments_applicationId - The application ID.

$sel:environmentId:ListDeployments', listDeployments_environmentId - The environment ID.

Request Lenses

listDeployments_nextToken :: Lens' ListDeployments (Maybe Text)

A token to start the list. Use this token to get the next set of results.

listDeployments_maxResults :: Lens' ListDeployments (Maybe Natural)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

listDeployments_applicationId :: Lens' ListDeployments Text

The application ID.

listDeployments_environmentId :: Lens' ListDeployments Text

The environment ID.

Destructuring the Response

data ListDeploymentsResponse

See: newListDeploymentsResponse smart constructor.

Constructors

ListDeploymentsResponse'

Fields

items :: Maybe [DeploymentSummary]

The elements from this collection.

nextToken :: Maybe Text

The token for the next set of items to return. Use this token to get the next set of results.

httpStatus :: Int

The response's http status code.

newListDeploymentsResponse

Arguments

:: Int
$sel:httpStatus:ListDeploymentsResponse'

-> ListDeploymentsResponse

Create a value of ListDeploymentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:ListDeploymentsResponse', listDeploymentsResponse_items - The elements from this collection.

$sel:nextToken:ListDeployments', listDeploymentsResponse_nextToken - The token for the next set of items to return. Use this token to get the next set of results.

$sel:httpStatus:ListDeploymentsResponse', listDeploymentsResponse_httpStatus - The response's http status code.

Response Lenses

listDeploymentsResponse_items :: Lens' ListDeploymentsResponse (Maybe [DeploymentSummary])

The elements from this collection.

listDeploymentsResponse_nextToken :: Lens' ListDeploymentsResponse (Maybe Text)

The token for the next set of items to return. Use this token to get the next set of results.

listDeploymentsResponse_httpStatus :: Lens' ListDeploymentsResponse Int

The response's http status code.