Network.AWS.AppConfig.Types

Description

Service Configuration

defaultService :: Service

API version 2019-10-09 of the Amazon AppConfig SDK configuration.

Errors

_PayloadTooLargeException :: AsError a => Getting (First ServiceError) a ServiceError

The configuration size is too large.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request could not be processed because of conflict in the current state of the resource.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of hosted configuration versions exceeds the limit for the AppConfig configuration store. Delete one or more versions and try again.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

There was an internal failure in the AppConfig service.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource could not be found.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

DeploymentEventType

newtype DeploymentEventType

Constructors

DeploymentEventType'

Fields

fromDeploymentEventType :: Text

Bundled Patterns

pattern DeploymentEventType_BAKE_TIME_STARTED :: DeploymentEventType

pattern DeploymentEventType_DEPLOYMENT_COMPLETED :: DeploymentEventType

pattern DeploymentEventType_DEPLOYMENT_STARTED :: DeploymentEventType

pattern DeploymentEventType_PERCENTAGE_UPDATED :: DeploymentEventType

pattern DeploymentEventType_ROLLBACK_COMPLETED :: DeploymentEventType

pattern DeploymentEventType_ROLLBACK_STARTED :: DeploymentEventType

DeploymentState

newtype DeploymentState

Constructors

DeploymentState'

Fields

fromDeploymentState :: Text

Bundled Patterns

pattern DeploymentState_BAKING :: DeploymentState

pattern DeploymentState_COMPLETE :: DeploymentState

pattern DeploymentState_DEPLOYING :: DeploymentState

pattern DeploymentState_ROLLED_BACK :: DeploymentState

pattern DeploymentState_ROLLING_BACK :: DeploymentState

pattern DeploymentState_VALIDATING :: DeploymentState

EnvironmentState

newtype EnvironmentState

Constructors

EnvironmentState'

Fields

fromEnvironmentState :: Text

Bundled Patterns

pattern EnvironmentState_DEPLOYING :: EnvironmentState

pattern EnvironmentState_READY_FOR_DEPLOYMENT :: EnvironmentState

pattern EnvironmentState_ROLLED_BACK :: EnvironmentState

pattern EnvironmentState_ROLLING_BACK :: EnvironmentState

GrowthType

newtype GrowthType

Constructors

GrowthType'

Fields

fromGrowthType :: Text

Bundled Patterns

pattern GrowthType_EXPONENTIAL :: GrowthType

pattern GrowthType_LINEAR :: GrowthType

ReplicateTo

newtype ReplicateTo

Constructors

ReplicateTo'

Fields

fromReplicateTo :: Text

Bundled Patterns

pattern ReplicateTo_NONE :: ReplicateTo

pattern ReplicateTo_SSM_DOCUMENT :: ReplicateTo

TriggeredBy

newtype TriggeredBy

Constructors

TriggeredBy'

Fields

fromTriggeredBy :: Text

Bundled Patterns

pattern TriggeredBy_APPCONFIG :: TriggeredBy

pattern TriggeredBy_CLOUDWATCH_ALARM :: TriggeredBy

pattern TriggeredBy_INTERNAL_ERROR :: TriggeredBy

pattern TriggeredBy_USER :: TriggeredBy

ValidatorType

newtype ValidatorType

Constructors

ValidatorType'

Fields

fromValidatorType :: Text

Bundled Patterns

pattern ValidatorType_JSON_SCHEMA :: ValidatorType

pattern ValidatorType_LAMBDA :: ValidatorType

Application

data Application

See: newApplication smart constructor.

Constructors

Application'

Fields

name :: Maybe Text

The application name.

id :: Maybe Text

The application ID.

description :: Maybe Text

The description of the application.

newApplication :: Application

Create a value of Application with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Application', application_name - The application name.

$sel:id:Application', application_id - The application ID.

$sel:description:Application', application_description - The description of the application.

application_name :: Lens' Application (Maybe Text)

The application name.

application_id :: Lens' Application (Maybe Text)

The application ID.

application_description :: Lens' Application (Maybe Text)

The description of the application.

ConfigurationProfile

data ConfigurationProfile

See: newConfigurationProfile smart constructor.

Constructors

ConfigurationProfile'

Fields

retrievalRoleArn :: Maybe Text

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

validators :: Maybe [Validator]

A list of methods for validating the configuration.

locationUri :: Maybe Text

The URI location of the configuration.

applicationId :: Maybe Text

The application ID.

name :: Maybe Text

The name of the configuration profile.

id :: Maybe Text

The configuration profile ID.

description :: Maybe Text

The configuration profile description.

newConfigurationProfile :: ConfigurationProfile

Create a value of ConfigurationProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retrievalRoleArn:ConfigurationProfile', configurationProfile_retrievalRoleArn - The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

$sel:validators:ConfigurationProfile', configurationProfile_validators - A list of methods for validating the configuration.

$sel:locationUri:ConfigurationProfile', configurationProfile_locationUri - The URI location of the configuration.

$sel:applicationId:ConfigurationProfile', configurationProfile_applicationId - The application ID.

$sel:name:ConfigurationProfile', configurationProfile_name - The name of the configuration profile.

$sel:id:ConfigurationProfile', configurationProfile_id - The configuration profile ID.

$sel:description:ConfigurationProfile', configurationProfile_description - The configuration profile description.

configurationProfile_retrievalRoleArn :: Lens' ConfigurationProfile (Maybe Text)

The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

configurationProfile_validators :: Lens' ConfigurationProfile (Maybe [Validator])

A list of methods for validating the configuration.

configurationProfile_locationUri :: Lens' ConfigurationProfile (Maybe Text)

The URI location of the configuration.

configurationProfile_applicationId :: Lens' ConfigurationProfile (Maybe Text)

The application ID.

configurationProfile_name :: Lens' ConfigurationProfile (Maybe Text)

The name of the configuration profile.

configurationProfile_id :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile ID.

configurationProfile_description :: Lens' ConfigurationProfile (Maybe Text)

The configuration profile description.

ConfigurationProfileSummary

data ConfigurationProfileSummary

A summary of a configuration profile.

See: newConfigurationProfileSummary smart constructor.

Constructors

ConfigurationProfileSummary'

Fields

locationUri :: Maybe Text

The URI location of the configuration.

applicationId :: Maybe Text

The application ID.

name :: Maybe Text

The name of the configuration profile.

id :: Maybe Text

The ID of the configuration profile.

validatorTypes :: Maybe [ValidatorType]

The types of validators in the configuration profile.

newConfigurationProfileSummary :: ConfigurationProfileSummary

Create a value of ConfigurationProfileSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationUri:ConfigurationProfileSummary', configurationProfileSummary_locationUri - The URI location of the configuration.

$sel:applicationId:ConfigurationProfileSummary', configurationProfileSummary_applicationId - The application ID.

$sel:name:ConfigurationProfileSummary', configurationProfileSummary_name - The name of the configuration profile.

$sel:id:ConfigurationProfileSummary', configurationProfileSummary_id - The ID of the configuration profile.

$sel:validatorTypes:ConfigurationProfileSummary', configurationProfileSummary_validatorTypes - The types of validators in the configuration profile.

configurationProfileSummary_locationUri :: Lens' ConfigurationProfileSummary (Maybe Text)

The URI location of the configuration.

configurationProfileSummary_applicationId :: Lens' ConfigurationProfileSummary (Maybe Text)

The application ID.

configurationProfileSummary_name :: Lens' ConfigurationProfileSummary (Maybe Text)

The name of the configuration profile.

configurationProfileSummary_id :: Lens' ConfigurationProfileSummary (Maybe Text)

The ID of the configuration profile.

configurationProfileSummary_validatorTypes :: Lens' ConfigurationProfileSummary (Maybe [ValidatorType])

The types of validators in the configuration profile.

Deployment

data Deployment

See: newDeployment smart constructor.

Constructors

Deployment'

Fields

growthFactor :: Maybe Double

The percentage of targets to receive a deployed configuration during each interval.

configurationName :: Maybe Text

The name of the configuration.

state :: Maybe DeploymentState

The state of the deployment.

deploymentStrategyId :: Maybe Text

The ID of the deployment strategy that was deployed.

deploymentNumber :: Maybe Int

The sequence number of the deployment.

configurationVersion :: Maybe Text

The configuration version that was deployed.

eventLog :: Maybe [DeploymentEvent]

A list containing all events related to a deployment. The most recent events are displayed first.

percentageComplete :: Maybe Double

The percentage of targets for which the deployment is available.

startedAt :: Maybe POSIX

The time the deployment started.

applicationId :: Maybe Text

The ID of the application that was deployed.

deploymentDurationInMinutes :: Maybe Natural

Total amount of time the deployment lasted.

environmentId :: Maybe Text

The ID of the environment that was deployed.

completedAt :: Maybe POSIX

The time the deployment completed.

configurationLocationUri :: Maybe Text

Information about the source location of the configuration.

finalBakeTimeInMinutes :: Maybe Natural

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

description :: Maybe Text

The description of the deployment.

configurationProfileId :: Maybe Text

The ID of the configuration profile that was deployed.

growthType :: Maybe GrowthType

The algorithm used to define how percentage grew over time.

newDeployment :: Deployment

Create a value of Deployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:growthFactor:Deployment', deployment_growthFactor - The percentage of targets to receive a deployed configuration during each interval.

$sel:configurationName:Deployment', deployment_configurationName - The name of the configuration.

$sel:state:Deployment', deployment_state - The state of the deployment.

$sel:deploymentStrategyId:Deployment', deployment_deploymentStrategyId - The ID of the deployment strategy that was deployed.

$sel:deploymentNumber:Deployment', deployment_deploymentNumber - The sequence number of the deployment.

$sel:configurationVersion:Deployment', deployment_configurationVersion - The configuration version that was deployed.

$sel:eventLog:Deployment', deployment_eventLog - A list containing all events related to a deployment. The most recent events are displayed first.

$sel:percentageComplete:Deployment', deployment_percentageComplete - The percentage of targets for which the deployment is available.

$sel:startedAt:Deployment', deployment_startedAt - The time the deployment started.

$sel:applicationId:Deployment', deployment_applicationId - The ID of the application that was deployed.

$sel:deploymentDurationInMinutes:Deployment', deployment_deploymentDurationInMinutes - Total amount of time the deployment lasted.

$sel:environmentId:Deployment', deployment_environmentId - The ID of the environment that was deployed.

$sel:completedAt:Deployment', deployment_completedAt - The time the deployment completed.

$sel:configurationLocationUri:Deployment', deployment_configurationLocationUri - Information about the source location of the configuration.

$sel:finalBakeTimeInMinutes:Deployment', deployment_finalBakeTimeInMinutes - The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

$sel:description:Deployment', deployment_description - The description of the deployment.

$sel:configurationProfileId:Deployment', deployment_configurationProfileId - The ID of the configuration profile that was deployed.

$sel:growthType:Deployment', deployment_growthType - The algorithm used to define how percentage grew over time.

deployment_growthFactor :: Lens' Deployment (Maybe Double)

The percentage of targets to receive a deployed configuration during each interval.

deployment_configurationName :: Lens' Deployment (Maybe Text)

The name of the configuration.

deployment_state :: Lens' Deployment (Maybe DeploymentState)

The state of the deployment.

deployment_deploymentStrategyId :: Lens' Deployment (Maybe Text)

The ID of the deployment strategy that was deployed.

deployment_deploymentNumber :: Lens' Deployment (Maybe Int)

The sequence number of the deployment.

deployment_configurationVersion :: Lens' Deployment (Maybe Text)

The configuration version that was deployed.

deployment_eventLog :: Lens' Deployment (Maybe [DeploymentEvent])

A list containing all events related to a deployment. The most recent events are displayed first.

deployment_percentageComplete :: Lens' Deployment (Maybe Double)

The percentage of targets for which the deployment is available.

deployment_startedAt :: Lens' Deployment (Maybe UTCTime)

The time the deployment started.

deployment_applicationId :: Lens' Deployment (Maybe Text)

The ID of the application that was deployed.

deployment_deploymentDurationInMinutes :: Lens' Deployment (Maybe Natural)

Total amount of time the deployment lasted.

deployment_environmentId :: Lens' Deployment (Maybe Text)

The ID of the environment that was deployed.

deployment_completedAt :: Lens' Deployment (Maybe UTCTime)

The time the deployment completed.

deployment_configurationLocationUri :: Lens' Deployment (Maybe Text)

Information about the source location of the configuration.

deployment_finalBakeTimeInMinutes :: Lens' Deployment (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deployment_description :: Lens' Deployment (Maybe Text)

The description of the deployment.

deployment_configurationProfileId :: Lens' Deployment (Maybe Text)

The ID of the configuration profile that was deployed.

deployment_growthType :: Lens' Deployment (Maybe GrowthType)

The algorithm used to define how percentage grew over time.

DeploymentEvent

data DeploymentEvent

An object that describes a deployment event.

See: newDeploymentEvent smart constructor.

Constructors

DeploymentEvent'

Fields

triggeredBy :: Maybe TriggeredBy

The entity that triggered the deployment event. Events can be triggered by a user, AWS AppConfig, an Amazon CloudWatch alarm, or an internal error.

occurredAt :: Maybe POSIX

The date and time the event occurred.

eventType :: Maybe DeploymentEventType

The type of deployment event. Deployment event types include the start, stop, or completion of a deployment; a percentage update; the start or stop of a bake period; the start or completion of a rollback.

description :: Maybe Text

A description of the deployment event. Descriptions include, but are not limited to, the user account or the CloudWatch alarm ARN that initiated a rollback, the percentage of hosts that received the deployment, or in the case of an internal error, a recommendation to attempt a new deployment.

newDeploymentEvent :: DeploymentEvent

Create a value of DeploymentEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:triggeredBy:DeploymentEvent', deploymentEvent_triggeredBy - The entity that triggered the deployment event. Events can be triggered by a user, AWS AppConfig, an Amazon CloudWatch alarm, or an internal error.

$sel:occurredAt:DeploymentEvent', deploymentEvent_occurredAt - The date and time the event occurred.

$sel:eventType:DeploymentEvent', deploymentEvent_eventType - The type of deployment event. Deployment event types include the start, stop, or completion of a deployment; a percentage update; the start or stop of a bake period; the start or completion of a rollback.

$sel:description:DeploymentEvent', deploymentEvent_description - A description of the deployment event. Descriptions include, but are not limited to, the user account or the CloudWatch alarm ARN that initiated a rollback, the percentage of hosts that received the deployment, or in the case of an internal error, a recommendation to attempt a new deployment.

deploymentEvent_triggeredBy :: Lens' DeploymentEvent (Maybe TriggeredBy)

The entity that triggered the deployment event. Events can be triggered by a user, AWS AppConfig, an Amazon CloudWatch alarm, or an internal error.

deploymentEvent_occurredAt :: Lens' DeploymentEvent (Maybe UTCTime)

The date and time the event occurred.

deploymentEvent_eventType :: Lens' DeploymentEvent (Maybe DeploymentEventType)

The type of deployment event. Deployment event types include the start, stop, or completion of a deployment; a percentage update; the start or stop of a bake period; the start or completion of a rollback.

deploymentEvent_description :: Lens' DeploymentEvent (Maybe Text)

A description of the deployment event. Descriptions include, but are not limited to, the user account or the CloudWatch alarm ARN that initiated a rollback, the percentage of hosts that received the deployment, or in the case of an internal error, a recommendation to attempt a new deployment.

DeploymentStrategy

data DeploymentStrategy

See: newDeploymentStrategy smart constructor.

Constructors

DeploymentStrategy'

Fields

growthFactor :: Maybe Double

The percentage of targets that received a deployed configuration during each interval.

replicateTo :: Maybe ReplicateTo

Save the deployment strategy to a Systems Manager (SSM) document.

name :: Maybe Text

The name of the deployment strategy.

id :: Maybe Text

The deployment strategy ID.

deploymentDurationInMinutes :: Maybe Natural

Total amount of time the deployment lasted.

finalBakeTimeInMinutes :: Maybe Natural

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

description :: Maybe Text

The description of the deployment strategy.

growthType :: Maybe GrowthType

The algorithm used to define how percentage grew over time.

newDeploymentStrategy :: DeploymentStrategy

Create a value of DeploymentStrategy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:growthFactor:DeploymentStrategy', deploymentStrategy_growthFactor - The percentage of targets that received a deployed configuration during each interval.

$sel:replicateTo:DeploymentStrategy', deploymentStrategy_replicateTo - Save the deployment strategy to a Systems Manager (SSM) document.

$sel:name:DeploymentStrategy', deploymentStrategy_name - The name of the deployment strategy.

$sel:id:DeploymentStrategy', deploymentStrategy_id - The deployment strategy ID.

$sel:deploymentDurationInMinutes:DeploymentStrategy', deploymentStrategy_deploymentDurationInMinutes - Total amount of time the deployment lasted.

$sel:finalBakeTimeInMinutes:DeploymentStrategy', deploymentStrategy_finalBakeTimeInMinutes - The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

$sel:description:DeploymentStrategy', deploymentStrategy_description - The description of the deployment strategy.

$sel:growthType:DeploymentStrategy', deploymentStrategy_growthType - The algorithm used to define how percentage grew over time.

deploymentStrategy_growthFactor :: Lens' DeploymentStrategy (Maybe Double)

The percentage of targets that received a deployed configuration during each interval.

deploymentStrategy_replicateTo :: Lens' DeploymentStrategy (Maybe ReplicateTo)

Save the deployment strategy to a Systems Manager (SSM) document.

deploymentStrategy_name :: Lens' DeploymentStrategy (Maybe Text)

The name of the deployment strategy.

deploymentStrategy_id :: Lens' DeploymentStrategy (Maybe Text)

The deployment strategy ID.

deploymentStrategy_deploymentDurationInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

Total amount of time the deployment lasted.

deploymentStrategy_finalBakeTimeInMinutes :: Lens' DeploymentStrategy (Maybe Natural)

The amount of time AppConfig monitored for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deploymentStrategy_description :: Lens' DeploymentStrategy (Maybe Text)

The description of the deployment strategy.

deploymentStrategy_growthType :: Lens' DeploymentStrategy (Maybe GrowthType)

The algorithm used to define how percentage grew over time.

DeploymentSummary

data DeploymentSummary

Information about the deployment.

See: newDeploymentSummary smart constructor.

Constructors

DeploymentSummary'

Fields

growthFactor :: Maybe Double

The percentage of targets to receive a deployed configuration during each interval.

configurationName :: Maybe Text

The name of the configuration.

state :: Maybe DeploymentState

The state of the deployment.

deploymentNumber :: Maybe Int

The sequence number of the deployment.

configurationVersion :: Maybe Text

The version of the configuration.

percentageComplete :: Maybe Double

The percentage of targets for which the deployment is available.

startedAt :: Maybe POSIX

Time the deployment started.

deploymentDurationInMinutes :: Maybe Natural

Total amount of time the deployment lasted.

completedAt :: Maybe POSIX

Time the deployment completed.

finalBakeTimeInMinutes :: Maybe Natural

The amount of time AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

growthType :: Maybe GrowthType

The algorithm used to define how percentage grows over time.

newDeploymentSummary :: DeploymentSummary

Create a value of DeploymentSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:growthFactor:DeploymentSummary', deploymentSummary_growthFactor - The percentage of targets to receive a deployed configuration during each interval.

$sel:configurationName:DeploymentSummary', deploymentSummary_configurationName - The name of the configuration.

$sel:state:DeploymentSummary', deploymentSummary_state - The state of the deployment.

$sel:deploymentNumber:DeploymentSummary', deploymentSummary_deploymentNumber - The sequence number of the deployment.

$sel:configurationVersion:DeploymentSummary', deploymentSummary_configurationVersion - The version of the configuration.

$sel:percentageComplete:DeploymentSummary', deploymentSummary_percentageComplete - The percentage of targets for which the deployment is available.

$sel:startedAt:DeploymentSummary', deploymentSummary_startedAt - Time the deployment started.

$sel:deploymentDurationInMinutes:DeploymentSummary', deploymentSummary_deploymentDurationInMinutes - Total amount of time the deployment lasted.

$sel:completedAt:DeploymentSummary', deploymentSummary_completedAt - Time the deployment completed.

$sel:finalBakeTimeInMinutes:DeploymentSummary', deploymentSummary_finalBakeTimeInMinutes - The amount of time AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

$sel:growthType:DeploymentSummary', deploymentSummary_growthType - The algorithm used to define how percentage grows over time.

deploymentSummary_growthFactor :: Lens' DeploymentSummary (Maybe Double)

The percentage of targets to receive a deployed configuration during each interval.

deploymentSummary_configurationName :: Lens' DeploymentSummary (Maybe Text)

The name of the configuration.

deploymentSummary_state :: Lens' DeploymentSummary (Maybe DeploymentState)

The state of the deployment.

deploymentSummary_deploymentNumber :: Lens' DeploymentSummary (Maybe Int)

The sequence number of the deployment.

deploymentSummary_configurationVersion :: Lens' DeploymentSummary (Maybe Text)

The version of the configuration.

deploymentSummary_percentageComplete :: Lens' DeploymentSummary (Maybe Double)

The percentage of targets for which the deployment is available.

deploymentSummary_startedAt :: Lens' DeploymentSummary (Maybe UTCTime)

Time the deployment started.

deploymentSummary_deploymentDurationInMinutes :: Lens' DeploymentSummary (Maybe Natural)

Total amount of time the deployment lasted.

deploymentSummary_completedAt :: Lens' DeploymentSummary (Maybe UTCTime)

Time the deployment completed.

deploymentSummary_finalBakeTimeInMinutes :: Lens' DeploymentSummary (Maybe Natural)

The amount of time AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic roll back.

deploymentSummary_growthType :: Lens' DeploymentSummary (Maybe GrowthType)

The algorithm used to define how percentage grows over time.

Environment

data Environment

See: newEnvironment smart constructor.

Constructors

Environment'

Fields

state :: Maybe EnvironmentState

The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

monitors :: Maybe [Monitor]

Amazon CloudWatch alarms monitored during the deployment.

applicationId :: Maybe Text

The application ID.

name :: Maybe Text

The name of the environment.

id :: Maybe Text

The environment ID.

description :: Maybe Text

The description of the environment.

newEnvironment :: Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Environment', environment_state - The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

$sel:monitors:Environment', environment_monitors - Amazon CloudWatch alarms monitored during the deployment.

$sel:applicationId:Environment', environment_applicationId - The application ID.

$sel:name:Environment', environment_name - The name of the environment.

$sel:id:Environment', environment_id - The environment ID.

$sel:description:Environment', environment_description - The description of the environment.

environment_state :: Lens' Environment (Maybe EnvironmentState)

The state of the environment. An environment can be in one of the following states: READY_FOR_DEPLOYMENT, DEPLOYING, ROLLING_BACK, or ROLLED_BACK

environment_monitors :: Lens' Environment (Maybe [Monitor])

Amazon CloudWatch alarms monitored during the deployment.

environment_applicationId :: Lens' Environment (Maybe Text)

The application ID.

environment_name :: Lens' Environment (Maybe Text)

The name of the environment.

environment_id :: Lens' Environment (Maybe Text)

The environment ID.

environment_description :: Lens' Environment (Maybe Text)

The description of the environment.

HostedConfigurationVersion

data HostedConfigurationVersion

See: newHostedConfigurationVersion smart constructor.

Constructors

HostedConfigurationVersion'

Fields

content :: Maybe (Sensitive ByteString)

The content of the configuration or the configuration data.

versionNumber :: Maybe Int

The configuration version.

applicationId :: Maybe Text

The application ID.

description :: Maybe Text

A description of the configuration.

configurationProfileId :: Maybe Text

The configuration profile ID.

contentType :: Maybe Text

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

newHostedConfigurationVersion :: HostedConfigurationVersion

Create a value of HostedConfigurationVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:content:HostedConfigurationVersion', hostedConfigurationVersion_content - The content of the configuration or the configuration data.

$sel:versionNumber:HostedConfigurationVersion', hostedConfigurationVersion_versionNumber - The configuration version.

$sel:applicationId:HostedConfigurationVersion', hostedConfigurationVersion_applicationId - The application ID.

$sel:description:HostedConfigurationVersion', hostedConfigurationVersion_description - A description of the configuration.

$sel:configurationProfileId:HostedConfigurationVersion', hostedConfigurationVersion_configurationProfileId - The configuration profile ID.

$sel:contentType:HostedConfigurationVersion', hostedConfigurationVersion_contentType - A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

hostedConfigurationVersion_content :: Lens' HostedConfigurationVersion (Maybe ByteString)

The content of the configuration or the configuration data.

hostedConfigurationVersion_versionNumber :: Lens' HostedConfigurationVersion (Maybe Int)

The configuration version.

hostedConfigurationVersion_applicationId :: Lens' HostedConfigurationVersion (Maybe Text)

The application ID.

hostedConfigurationVersion_description :: Lens' HostedConfigurationVersion (Maybe Text)

A description of the configuration.

hostedConfigurationVersion_configurationProfileId :: Lens' HostedConfigurationVersion (Maybe Text)

The configuration profile ID.

hostedConfigurationVersion_contentType :: Lens' HostedConfigurationVersion (Maybe Text)

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

HostedConfigurationVersionSummary

data HostedConfigurationVersionSummary

Information about the configuration.

See: newHostedConfigurationVersionSummary smart constructor.

Constructors

HostedConfigurationVersionSummary'

Fields

versionNumber :: Maybe Int

The configuration version.

applicationId :: Maybe Text

The application ID.

description :: Maybe Text

A description of the configuration.

configurationProfileId :: Maybe Text

The configuration profile ID.

contentType :: Maybe Text

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

newHostedConfigurationVersionSummary :: HostedConfigurationVersionSummary

Create a value of HostedConfigurationVersionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionNumber:HostedConfigurationVersionSummary', hostedConfigurationVersionSummary_versionNumber - The configuration version.

$sel:applicationId:HostedConfigurationVersionSummary', hostedConfigurationVersionSummary_applicationId - The application ID.

$sel:description:HostedConfigurationVersionSummary', hostedConfigurationVersionSummary_description - A description of the configuration.

$sel:configurationProfileId:HostedConfigurationVersionSummary', hostedConfigurationVersionSummary_configurationProfileId - The configuration profile ID.

$sel:contentType:HostedConfigurationVersionSummary', hostedConfigurationVersionSummary_contentType - A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

hostedConfigurationVersionSummary_versionNumber :: Lens' HostedConfigurationVersionSummary (Maybe Int)

The configuration version.

hostedConfigurationVersionSummary_applicationId :: Lens' HostedConfigurationVersionSummary (Maybe Text)

The application ID.

hostedConfigurationVersionSummary_description :: Lens' HostedConfigurationVersionSummary (Maybe Text)

A description of the configuration.

hostedConfigurationVersionSummary_configurationProfileId :: Lens' HostedConfigurationVersionSummary (Maybe Text)

The configuration profile ID.

hostedConfigurationVersionSummary_contentType :: Lens' HostedConfigurationVersionSummary (Maybe Text)

A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

Monitor

data Monitor

Amazon CloudWatch alarms to monitor during the deployment process.

See: newMonitor smart constructor.

Constructors

Monitor'

Fields

alarmRoleArn :: Maybe Text

ARN of an IAM role for AppConfig to monitor AlarmArn.

alarmArn :: Maybe Text

ARN of the Amazon CloudWatch alarm.

newMonitor :: Monitor

Create a value of Monitor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alarmRoleArn:Monitor', monitor_alarmRoleArn - ARN of an IAM role for AppConfig to monitor AlarmArn.

$sel:alarmArn:Monitor', monitor_alarmArn - ARN of the Amazon CloudWatch alarm.

monitor_alarmRoleArn :: Lens' Monitor (Maybe Text)

ARN of an IAM role for AppConfig to monitor AlarmArn.

monitor_alarmArn :: Lens' Monitor (Maybe Text)

ARN of the Amazon CloudWatch alarm.

Validator

data Validator

A validator provides a syntactic or semantic check to ensure the configuration you want to deploy functions as intended. To validate your application configuration data, you provide a schema or a Lambda function that runs against the configuration. The configuration deployment or update can only proceed when the configuration data is valid.

See: newValidator smart constructor.

Constructors

Validator'

Fields

type' :: ValidatorType

AppConfig supports validators of type JSON_SCHEMA and LAMBDA

content :: Sensitive Text

Either the JSON Schema content or the Amazon Resource Name (ARN) of an AWS Lambda function.

newValidator

Arguments

:: ValidatorType
$sel:type':Validator'

-> Text
$sel:content:Validator'

-> Validator

Create a value of Validator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':Validator', validator_type - AppConfig supports validators of type JSON_SCHEMA and LAMBDA

$sel:content:Validator', validator_content - Either the JSON Schema content or the Amazon Resource Name (ARN) of an AWS Lambda function.

validator_type :: Lens' Validator ValidatorType

AppConfig supports validators of type JSON_SCHEMA and LAMBDA

validator_content :: Lens' Validator Text

Either the JSON Schema content or the Amazon Resource Name (ARN) of an AWS Lambda function.