Network.AWS.AppFlow.DescribeFlow

Description

Provides a description of the specified flow.

Creating a Request

data DescribeFlow

See: newDescribeFlow smart constructor.

Constructors

DescribeFlow'

Fields

flowName :: Text

The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

newDescribeFlow

Arguments

:: Text
$sel:flowName:DescribeFlow'

-> DescribeFlow

Create a value of DescribeFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowName:DescribeFlow', describeFlow_flowName - The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

Request Lenses

describeFlow_flowName :: Lens' DescribeFlow Text

The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

Destructuring the Response

data DescribeFlowResponse

See: newDescribeFlowResponse smart constructor.

Constructors

DescribeFlowResponse'

Fields

lastUpdatedBy :: Maybe Text

Specifies the user name of the account that performed the most recent update.

flowArn :: Maybe Text

The flow's Amazon Resource Name (ARN).

lastUpdatedAt :: Maybe POSIX

Specifies when the flow was last updated.

createdAt :: Maybe POSIX

Specifies when the flow was created.

createdBy :: Maybe Text

The ARN of the user who created the flow.

tasks :: Maybe [Task]

A list of tasks that Amazon AppFlow performs while transferring the data in the flow run.

triggerConfig :: Maybe TriggerConfig

The trigger settings that determine how and when the flow runs.

flowName :: Maybe Text

The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

sourceFlowConfig :: Maybe SourceFlowConfig

The configuration that controls how Amazon AppFlow retrieves data from the source connector.

flowStatusMessage :: Maybe Text

Contains an error message if the flow status is in a suspended or error state. This applies only to scheduled or event-triggered flows.

kmsArn :: Maybe Text

The ARN (Amazon Resource Name) of the Key Management Service (KMS) key you provide for encryption. This is required if you do not want to use the Amazon AppFlow-managed KMS key. If you don't provide anything here, Amazon AppFlow uses the Amazon AppFlow-managed KMS key.

lastRunExecutionDetails :: Maybe ExecutionDetails

Describes the details of the most recent flow run.

flowStatus :: Maybe FlowStatus

Indicates the current status of the flow.

destinationFlowConfigList :: Maybe [DestinationFlowConfig]

The configuration that controls how Amazon AppFlow transfers data to the destination connector.

description :: Maybe Text

A description of the flow.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for your flow.

httpStatus :: Int

The response's http status code.

newDescribeFlowResponse

Arguments

:: Int
$sel:httpStatus:DescribeFlowResponse'

-> DescribeFlowResponse

Create a value of DescribeFlowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedBy:DescribeFlowResponse', describeFlowResponse_lastUpdatedBy - Specifies the user name of the account that performed the most recent update.

$sel:flowArn:DescribeFlowResponse', describeFlowResponse_flowArn - The flow's Amazon Resource Name (ARN).

$sel:lastUpdatedAt:DescribeFlowResponse', describeFlowResponse_lastUpdatedAt - Specifies when the flow was last updated.

$sel:createdAt:DescribeFlowResponse', describeFlowResponse_createdAt - Specifies when the flow was created.

$sel:createdBy:DescribeFlowResponse', describeFlowResponse_createdBy - The ARN of the user who created the flow.

$sel:tasks:DescribeFlowResponse', describeFlowResponse_tasks - A list of tasks that Amazon AppFlow performs while transferring the data in the flow run.

$sel:triggerConfig:DescribeFlowResponse', describeFlowResponse_triggerConfig - The trigger settings that determine how and when the flow runs.

$sel:flowName:DescribeFlow', describeFlowResponse_flowName - The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

$sel:sourceFlowConfig:DescribeFlowResponse', describeFlowResponse_sourceFlowConfig - The configuration that controls how Amazon AppFlow retrieves data from the source connector.

$sel:flowStatusMessage:DescribeFlowResponse', describeFlowResponse_flowStatusMessage - Contains an error message if the flow status is in a suspended or error state. This applies only to scheduled or event-triggered flows.

$sel:kmsArn:DescribeFlowResponse', describeFlowResponse_kmsArn - The ARN (Amazon Resource Name) of the Key Management Service (KMS) key you provide for encryption. This is required if you do not want to use the Amazon AppFlow-managed KMS key. If you don't provide anything here, Amazon AppFlow uses the Amazon AppFlow-managed KMS key.

$sel:lastRunExecutionDetails:DescribeFlowResponse', describeFlowResponse_lastRunExecutionDetails - Describes the details of the most recent flow run.

$sel:flowStatus:DescribeFlowResponse', describeFlowResponse_flowStatus - Indicates the current status of the flow.

$sel:destinationFlowConfigList:DescribeFlowResponse', describeFlowResponse_destinationFlowConfigList - The configuration that controls how Amazon AppFlow transfers data to the destination connector.

$sel:description:DescribeFlowResponse', describeFlowResponse_description - A description of the flow.

$sel:tags:DescribeFlowResponse', describeFlowResponse_tags - The tags used to organize, track, or control access for your flow.

$sel:httpStatus:DescribeFlowResponse', describeFlowResponse_httpStatus - The response's http status code.

Response Lenses

describeFlowResponse_lastUpdatedBy :: Lens' DescribeFlowResponse (Maybe Text)

Specifies the user name of the account that performed the most recent update.

describeFlowResponse_flowArn :: Lens' DescribeFlowResponse (Maybe Text)

The flow's Amazon Resource Name (ARN).

describeFlowResponse_lastUpdatedAt :: Lens' DescribeFlowResponse (Maybe UTCTime)

Specifies when the flow was last updated.

describeFlowResponse_createdAt :: Lens' DescribeFlowResponse (Maybe UTCTime)

Specifies when the flow was created.

describeFlowResponse_createdBy :: Lens' DescribeFlowResponse (Maybe Text)

The ARN of the user who created the flow.

describeFlowResponse_tasks :: Lens' DescribeFlowResponse (Maybe [Task])

A list of tasks that Amazon AppFlow performs while transferring the data in the flow run.

describeFlowResponse_triggerConfig :: Lens' DescribeFlowResponse (Maybe TriggerConfig)

The trigger settings that determine how and when the flow runs.

describeFlowResponse_flowName :: Lens' DescribeFlowResponse (Maybe Text)

The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

describeFlowResponse_sourceFlowConfig :: Lens' DescribeFlowResponse (Maybe SourceFlowConfig)

The configuration that controls how Amazon AppFlow retrieves data from the source connector.

describeFlowResponse_flowStatusMessage :: Lens' DescribeFlowResponse (Maybe Text)

Contains an error message if the flow status is in a suspended or error state. This applies only to scheduled or event-triggered flows.

describeFlowResponse_kmsArn :: Lens' DescribeFlowResponse (Maybe Text)

The ARN (Amazon Resource Name) of the Key Management Service (KMS) key you provide for encryption. This is required if you do not want to use the Amazon AppFlow-managed KMS key. If you don't provide anything here, Amazon AppFlow uses the Amazon AppFlow-managed KMS key.

describeFlowResponse_lastRunExecutionDetails :: Lens' DescribeFlowResponse (Maybe ExecutionDetails)

Describes the details of the most recent flow run.

describeFlowResponse_flowStatus :: Lens' DescribeFlowResponse (Maybe FlowStatus)

Indicates the current status of the flow.

describeFlowResponse_destinationFlowConfigList :: Lens' DescribeFlowResponse (Maybe [DestinationFlowConfig])

The configuration that controls how Amazon AppFlow transfers data to the destination connector.

describeFlowResponse_description :: Lens' DescribeFlowResponse (Maybe Text)

A description of the flow.

describeFlowResponse_tags :: Lens' DescribeFlowResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for your flow.

describeFlowResponse_httpStatus :: Lens' DescribeFlowResponse Int

The response's http status code.