Network.AWS.AppFlow.DescribeFlowExecutionRecords

Description

Fetches the execution history of the flow.

Creating a Request

data DescribeFlowExecutionRecords

See: newDescribeFlowExecutionRecords smart constructor.

Constructors

DescribeFlowExecutionRecords'

Fields

nextToken :: Maybe Text

The pagination token for the next page of data.

maxResults :: Maybe Natural

Specifies the maximum number of items that should be returned in the result set. The default for maxResults is 20 (for all paginated API operations).

flowName :: Text

The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

newDescribeFlowExecutionRecords

Arguments

:: Text
$sel:flowName:DescribeFlowExecutionRecords'

-> DescribeFlowExecutionRecords

Create a value of DescribeFlowExecutionRecords with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeFlowExecutionRecords', describeFlowExecutionRecords_nextToken - The pagination token for the next page of data.

$sel:maxResults:DescribeFlowExecutionRecords', describeFlowExecutionRecords_maxResults - Specifies the maximum number of items that should be returned in the result set. The default for maxResults is 20 (for all paginated API operations).

$sel:flowName:DescribeFlowExecutionRecords', describeFlowExecutionRecords_flowName - The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

Request Lenses

describeFlowExecutionRecords_nextToken :: Lens' DescribeFlowExecutionRecords (Maybe Text)

The pagination token for the next page of data.

describeFlowExecutionRecords_maxResults :: Lens' DescribeFlowExecutionRecords (Maybe Natural)

Specifies the maximum number of items that should be returned in the result set. The default for maxResults is 20 (for all paginated API operations).

describeFlowExecutionRecords_flowName :: Lens' DescribeFlowExecutionRecords Text

The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

Destructuring the Response

data DescribeFlowExecutionRecordsResponse

See: newDescribeFlowExecutionRecordsResponse smart constructor.

Constructors

DescribeFlowExecutionRecordsResponse'

Fields

flowExecutions :: Maybe [ExecutionRecord]

Returns a list of all instances when this flow was run.

nextToken :: Maybe Text

The pagination token for the next page of data.

httpStatus :: Int

The response's http status code.

newDescribeFlowExecutionRecordsResponse

Arguments

:: Int
$sel:httpStatus:DescribeFlowExecutionRecordsResponse'

-> DescribeFlowExecutionRecordsResponse

Create a value of DescribeFlowExecutionRecordsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowExecutions:DescribeFlowExecutionRecordsResponse', describeFlowExecutionRecordsResponse_flowExecutions - Returns a list of all instances when this flow was run.

$sel:nextToken:DescribeFlowExecutionRecords', describeFlowExecutionRecordsResponse_nextToken - The pagination token for the next page of data.

$sel:httpStatus:DescribeFlowExecutionRecordsResponse', describeFlowExecutionRecordsResponse_httpStatus - The response's http status code.

Response Lenses

describeFlowExecutionRecordsResponse_flowExecutions :: Lens' DescribeFlowExecutionRecordsResponse (Maybe [ExecutionRecord])

Returns a list of all instances when this flow was run.

describeFlowExecutionRecordsResponse_nextToken :: Lens' DescribeFlowExecutionRecordsResponse (Maybe Text)

The pagination token for the next page of data.

describeFlowExecutionRecordsResponse_httpStatus :: Lens' DescribeFlowExecutionRecordsResponse Int

The response's http status code.