Network.AWS.AppFlow.ListTagsForResource

Description

Retrieves the tags that are associated with a specified flow.

Creating a Request

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource'

Fields

resourceArn :: Text

The Amazon Resource Name (ARN) of the specified flow.

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon Resource Name (ARN) of the specified flow.

Request Lenses

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the specified flow.

Destructuring the Response

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse'

Fields

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for your flow.

httpStatus :: Int

The response's http status code.

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The tags used to organize, track, or control access for your flow.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

Response Lenses

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for your flow.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.