Network.AWS.AppFlow.Types.ConnectorConfiguration

Description

Documentation

data ConnectorConfiguration

The configuration settings related to a given connector.

See: newConnectorConfiguration smart constructor.

Constructors

ConnectorConfiguration'

Fields

isPrivateLinkEnabled :: Maybe Bool

Specifies if PrivateLink is enabled for that connector.

supportedTriggerTypes :: Maybe [TriggerType]

Specifies the supported trigger types for the flow.

canUseAsSource :: Maybe Bool

Specifies whether the connector can be used as a source.

connectorMetadata :: Maybe ConnectorMetadata

Specifies connector-specific metadata such as oAuthScopes, supportedRegions, privateLinkServiceUrl, and so on.

canUseAsDestination :: Maybe Bool

Specifies whether the connector can be used as a destination.

supportedSchedulingFrequencies :: Maybe [ScheduleFrequencyType]

Specifies the supported flow frequency for that connector.

supportedDestinationConnectors :: Maybe [ConnectorType]

Lists the connectors that are available for use as destinations.

isPrivateLinkEndpointUrlRequired :: Maybe Bool

Specifies if a PrivateLink endpoint URL is required.

newConnectorConfiguration :: ConnectorConfiguration

Create a value of ConnectorConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isPrivateLinkEnabled:ConnectorConfiguration', connectorConfiguration_isPrivateLinkEnabled - Specifies if PrivateLink is enabled for that connector.

$sel:supportedTriggerTypes:ConnectorConfiguration', connectorConfiguration_supportedTriggerTypes - Specifies the supported trigger types for the flow.

$sel:canUseAsSource:ConnectorConfiguration', connectorConfiguration_canUseAsSource - Specifies whether the connector can be used as a source.

$sel:connectorMetadata:ConnectorConfiguration', connectorConfiguration_connectorMetadata - Specifies connector-specific metadata such as oAuthScopes, supportedRegions, privateLinkServiceUrl, and so on.

$sel:canUseAsDestination:ConnectorConfiguration', connectorConfiguration_canUseAsDestination - Specifies whether the connector can be used as a destination.

$sel:supportedSchedulingFrequencies:ConnectorConfiguration', connectorConfiguration_supportedSchedulingFrequencies - Specifies the supported flow frequency for that connector.

$sel:supportedDestinationConnectors:ConnectorConfiguration', connectorConfiguration_supportedDestinationConnectors - Lists the connectors that are available for use as destinations.

$sel:isPrivateLinkEndpointUrlRequired:ConnectorConfiguration', connectorConfiguration_isPrivateLinkEndpointUrlRequired - Specifies if a PrivateLink endpoint URL is required.

connectorConfiguration_isPrivateLinkEnabled :: Lens' ConnectorConfiguration (Maybe Bool)

Specifies if PrivateLink is enabled for that connector.

connectorConfiguration_supportedTriggerTypes :: Lens' ConnectorConfiguration (Maybe [TriggerType])

Specifies the supported trigger types for the flow.

connectorConfiguration_canUseAsSource :: Lens' ConnectorConfiguration (Maybe Bool)

Specifies whether the connector can be used as a source.

connectorConfiguration_connectorMetadata :: Lens' ConnectorConfiguration (Maybe ConnectorMetadata)

Specifies connector-specific metadata such as oAuthScopes, supportedRegions, privateLinkServiceUrl, and so on.

connectorConfiguration_canUseAsDestination :: Lens' ConnectorConfiguration (Maybe Bool)

Specifies whether the connector can be used as a destination.

connectorConfiguration_supportedSchedulingFrequencies :: Lens' ConnectorConfiguration (Maybe [ScheduleFrequencyType])

Specifies the supported flow frequency for that connector.

connectorConfiguration_supportedDestinationConnectors :: Lens' ConnectorConfiguration (Maybe [ConnectorType])

Lists the connectors that are available for use as destinations.

connectorConfiguration_isPrivateLinkEndpointUrlRequired :: Lens' ConnectorConfiguration (Maybe Bool)

Specifies if a PrivateLink endpoint URL is required.