Network.AWS.AppFlow.Types.ConnectorMetadata

Description

Documentation

data ConnectorMetadata

A structure to specify connector-specific metadata such as oAuthScopes, supportedRegions, privateLinkServiceUrl, and so on.

See: newConnectorMetadata smart constructor.

Constructors

ConnectorMetadata'

Fields

upsolver :: Maybe UpsolverMetadata

The connector metadata specific to Upsolver.

snowflake :: Maybe SnowflakeMetadata

The connector metadata specific to Snowflake.

honeycode :: Maybe HoneycodeMetadata

The connector metadata specific to Amazon Honeycode.

serviceNow :: Maybe ServiceNowMetadata

The connector metadata specific to ServiceNow.

dynatrace :: Maybe DynatraceMetadata

The connector metadata specific to Dynatrace.

marketo :: Maybe MarketoMetadata

The connector metadata specific to Marketo.

slack :: Maybe SlackMetadata

The connector metadata specific to Slack.

singular :: Maybe SingularMetadata

The connector metadata specific to Singular.

inforNexus :: Maybe InforNexusMetadata

The connector metadata specific to Infor Nexus.

amplitude :: Maybe AmplitudeMetadata

The connector metadata specific to Amplitude.

customerProfiles :: Maybe CustomerProfilesMetadata

The connector metadata specific to Amazon Connect Customer Profiles.

datadog :: Maybe DatadogMetadata

The connector metadata specific to Datadog.

googleAnalytics :: Maybe GoogleAnalyticsMetadata

The connector metadata specific to Google Analytics.

sAPOData :: Maybe SAPODataMetadata

salesforce :: Maybe SalesforceMetadata

The connector metadata specific to Salesforce.

zendesk :: Maybe ZendeskMetadata

The connector metadata specific to Zendesk.

s3 :: Maybe S3Metadata

The connector metadata specific to Amazon S3.

eventBridge :: Maybe EventBridgeMetadata

The connector metadata specific to Amazon EventBridge.

trendmicro :: Maybe TrendmicroMetadata

The connector metadata specific to Trend Micro.

redshift :: Maybe RedshiftMetadata

The connector metadata specific to Amazon Redshift.

veeva :: Maybe VeevaMetadata

The connector metadata specific to Veeva.

newConnectorMetadata :: ConnectorMetadata

Create a value of ConnectorMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:upsolver:ConnectorMetadata', connectorMetadata_upsolver - The connector metadata specific to Upsolver.

$sel:snowflake:ConnectorMetadata', connectorMetadata_snowflake - The connector metadata specific to Snowflake.

$sel:honeycode:ConnectorMetadata', connectorMetadata_honeycode - The connector metadata specific to Amazon Honeycode.

$sel:serviceNow:ConnectorMetadata', connectorMetadata_serviceNow - The connector metadata specific to ServiceNow.

$sel:dynatrace:ConnectorMetadata', connectorMetadata_dynatrace - The connector metadata specific to Dynatrace.

$sel:marketo:ConnectorMetadata', connectorMetadata_marketo - The connector metadata specific to Marketo.

$sel:slack:ConnectorMetadata', connectorMetadata_slack - The connector metadata specific to Slack.

$sel:singular:ConnectorMetadata', connectorMetadata_singular - The connector metadata specific to Singular.

$sel:inforNexus:ConnectorMetadata', connectorMetadata_inforNexus - The connector metadata specific to Infor Nexus.

$sel:amplitude:ConnectorMetadata', connectorMetadata_amplitude - The connector metadata specific to Amplitude.

$sel:customerProfiles:ConnectorMetadata', connectorMetadata_customerProfiles - The connector metadata specific to Amazon Connect Customer Profiles.

$sel:datadog:ConnectorMetadata', connectorMetadata_datadog - The connector metadata specific to Datadog.

$sel:googleAnalytics:ConnectorMetadata', connectorMetadata_googleAnalytics - The connector metadata specific to Google Analytics.

$sel:sAPOData:ConnectorMetadata', connectorMetadata_sAPOData - Undocumented member.

$sel:salesforce:ConnectorMetadata', connectorMetadata_salesforce - The connector metadata specific to Salesforce.

$sel:zendesk:ConnectorMetadata', connectorMetadata_zendesk - The connector metadata specific to Zendesk.

$sel:s3:ConnectorMetadata', connectorMetadata_s3 - The connector metadata specific to Amazon S3.

$sel:eventBridge:ConnectorMetadata', connectorMetadata_eventBridge - The connector metadata specific to Amazon EventBridge.

$sel:trendmicro:ConnectorMetadata', connectorMetadata_trendmicro - The connector metadata specific to Trend Micro.

$sel:redshift:ConnectorMetadata', connectorMetadata_redshift - The connector metadata specific to Amazon Redshift.

$sel:veeva:ConnectorMetadata', connectorMetadata_veeva - The connector metadata specific to Veeva.

connectorMetadata_upsolver :: Lens' ConnectorMetadata (Maybe UpsolverMetadata)

The connector metadata specific to Upsolver.

connectorMetadata_snowflake :: Lens' ConnectorMetadata (Maybe SnowflakeMetadata)

The connector metadata specific to Snowflake.

connectorMetadata_honeycode :: Lens' ConnectorMetadata (Maybe HoneycodeMetadata)

The connector metadata specific to Amazon Honeycode.

connectorMetadata_serviceNow :: Lens' ConnectorMetadata (Maybe ServiceNowMetadata)

The connector metadata specific to ServiceNow.

connectorMetadata_dynatrace :: Lens' ConnectorMetadata (Maybe DynatraceMetadata)

The connector metadata specific to Dynatrace.

connectorMetadata_marketo :: Lens' ConnectorMetadata (Maybe MarketoMetadata)

The connector metadata specific to Marketo.

connectorMetadata_slack :: Lens' ConnectorMetadata (Maybe SlackMetadata)

The connector metadata specific to Slack.

connectorMetadata_singular :: Lens' ConnectorMetadata (Maybe SingularMetadata)

The connector metadata specific to Singular.

connectorMetadata_inforNexus :: Lens' ConnectorMetadata (Maybe InforNexusMetadata)

The connector metadata specific to Infor Nexus.

connectorMetadata_amplitude :: Lens' ConnectorMetadata (Maybe AmplitudeMetadata)

The connector metadata specific to Amplitude.

connectorMetadata_customerProfiles :: Lens' ConnectorMetadata (Maybe CustomerProfilesMetadata)

The connector metadata specific to Amazon Connect Customer Profiles.

connectorMetadata_datadog :: Lens' ConnectorMetadata (Maybe DatadogMetadata)

The connector metadata specific to Datadog.

connectorMetadata_googleAnalytics :: Lens' ConnectorMetadata (Maybe GoogleAnalyticsMetadata)

The connector metadata specific to Google Analytics.

connectorMetadata_sAPOData :: Lens' ConnectorMetadata (Maybe SAPODataMetadata)

Undocumented member.

connectorMetadata_salesforce :: Lens' ConnectorMetadata (Maybe SalesforceMetadata)

The connector metadata specific to Salesforce.

connectorMetadata_zendesk :: Lens' ConnectorMetadata (Maybe ZendeskMetadata)

The connector metadata specific to Zendesk.

connectorMetadata_s3 :: Lens' ConnectorMetadata (Maybe S3Metadata)

The connector metadata specific to Amazon S3.

connectorMetadata_eventBridge :: Lens' ConnectorMetadata (Maybe EventBridgeMetadata)

The connector metadata specific to Amazon EventBridge.

connectorMetadata_trendmicro :: Lens' ConnectorMetadata (Maybe TrendmicroMetadata)

The connector metadata specific to Trend Micro.

connectorMetadata_redshift :: Lens' ConnectorMetadata (Maybe RedshiftMetadata)

The connector metadata specific to Amazon Redshift.

connectorMetadata_veeva :: Lens' ConnectorMetadata (Maybe VeevaMetadata)

The connector metadata specific to Veeva.