Network.AWS.AppFlow.Types.ConnectorOperator

Description

Documentation

data ConnectorOperator

The operation to be performed on the provided source fields.

See: newConnectorOperator smart constructor.

Constructors

ConnectorOperator'

Fields

serviceNow :: Maybe ServiceNowConnectorOperator

The operation to be performed on the provided ServiceNow source fields.

dynatrace :: Maybe DynatraceConnectorOperator

The operation to be performed on the provided Dynatrace source fields.

marketo :: Maybe MarketoConnectorOperator

The operation to be performed on the provided Marketo source fields.

slack :: Maybe SlackConnectorOperator

The operation to be performed on the provided Slack source fields.

singular :: Maybe SingularConnectorOperator

The operation to be performed on the provided Singular source fields.

inforNexus :: Maybe InforNexusConnectorOperator

The operation to be performed on the provided Infor Nexus source fields.

amplitude :: Maybe AmplitudeConnectorOperator

The operation to be performed on the provided Amplitude source fields.

datadog :: Maybe DatadogConnectorOperator

The operation to be performed on the provided Datadog source fields.

googleAnalytics :: Maybe GoogleAnalyticsConnectorOperator

The operation to be performed on the provided Google Analytics source fields.

sAPOData :: Maybe SAPODataConnectorOperator

The operation to be performed on the provided SAPOData source fields.

salesforce :: Maybe SalesforceConnectorOperator

The operation to be performed on the provided Salesforce source fields.

zendesk :: Maybe ZendeskConnectorOperator

The operation to be performed on the provided Zendesk source fields.

s3 :: Maybe S3ConnectorOperator

The operation to be performed on the provided Amazon S3 source fields.

trendmicro :: Maybe TrendmicroConnectorOperator

The operation to be performed on the provided Trend Micro source fields.

veeva :: Maybe VeevaConnectorOperator

The operation to be performed on the provided Veeva source fields.

newConnectorOperator :: ConnectorOperator

Create a value of ConnectorOperator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceNow:ConnectorOperator', connectorOperator_serviceNow - The operation to be performed on the provided ServiceNow source fields.

$sel:dynatrace:ConnectorOperator', connectorOperator_dynatrace - The operation to be performed on the provided Dynatrace source fields.

$sel:marketo:ConnectorOperator', connectorOperator_marketo - The operation to be performed on the provided Marketo source fields.

$sel:slack:ConnectorOperator', connectorOperator_slack - The operation to be performed on the provided Slack source fields.

$sel:singular:ConnectorOperator', connectorOperator_singular - The operation to be performed on the provided Singular source fields.

$sel:inforNexus:ConnectorOperator', connectorOperator_inforNexus - The operation to be performed on the provided Infor Nexus source fields.

$sel:amplitude:ConnectorOperator', connectorOperator_amplitude - The operation to be performed on the provided Amplitude source fields.

$sel:datadog:ConnectorOperator', connectorOperator_datadog - The operation to be performed on the provided Datadog source fields.

$sel:googleAnalytics:ConnectorOperator', connectorOperator_googleAnalytics - The operation to be performed on the provided Google Analytics source fields.

$sel:sAPOData:ConnectorOperator', connectorOperator_sAPOData - The operation to be performed on the provided SAPOData source fields.

$sel:salesforce:ConnectorOperator', connectorOperator_salesforce - The operation to be performed on the provided Salesforce source fields.

$sel:zendesk:ConnectorOperator', connectorOperator_zendesk - The operation to be performed on the provided Zendesk source fields.

$sel:s3:ConnectorOperator', connectorOperator_s3 - The operation to be performed on the provided Amazon S3 source fields.

$sel:trendmicro:ConnectorOperator', connectorOperator_trendmicro - The operation to be performed on the provided Trend Micro source fields.

$sel:veeva:ConnectorOperator', connectorOperator_veeva - The operation to be performed on the provided Veeva source fields.

connectorOperator_serviceNow :: Lens' ConnectorOperator (Maybe ServiceNowConnectorOperator)

The operation to be performed on the provided ServiceNow source fields.

connectorOperator_dynatrace :: Lens' ConnectorOperator (Maybe DynatraceConnectorOperator)

The operation to be performed on the provided Dynatrace source fields.

connectorOperator_marketo :: Lens' ConnectorOperator (Maybe MarketoConnectorOperator)

The operation to be performed on the provided Marketo source fields.

connectorOperator_slack :: Lens' ConnectorOperator (Maybe SlackConnectorOperator)

The operation to be performed on the provided Slack source fields.

connectorOperator_singular :: Lens' ConnectorOperator (Maybe SingularConnectorOperator)

The operation to be performed on the provided Singular source fields.

connectorOperator_inforNexus :: Lens' ConnectorOperator (Maybe InforNexusConnectorOperator)

The operation to be performed on the provided Infor Nexus source fields.

connectorOperator_amplitude :: Lens' ConnectorOperator (Maybe AmplitudeConnectorOperator)

The operation to be performed on the provided Amplitude source fields.

connectorOperator_datadog :: Lens' ConnectorOperator (Maybe DatadogConnectorOperator)

The operation to be performed on the provided Datadog source fields.

connectorOperator_googleAnalytics :: Lens' ConnectorOperator (Maybe GoogleAnalyticsConnectorOperator)

The operation to be performed on the provided Google Analytics source fields.

connectorOperator_sAPOData :: Lens' ConnectorOperator (Maybe SAPODataConnectorOperator)

The operation to be performed on the provided SAPOData source fields.

connectorOperator_salesforce :: Lens' ConnectorOperator (Maybe SalesforceConnectorOperator)

The operation to be performed on the provided Salesforce source fields.

connectorOperator_zendesk :: Lens' ConnectorOperator (Maybe ZendeskConnectorOperator)

The operation to be performed on the provided Zendesk source fields.

connectorOperator_s3 :: Lens' ConnectorOperator (Maybe S3ConnectorOperator)

The operation to be performed on the provided Amazon S3 source fields.

connectorOperator_trendmicro :: Lens' ConnectorOperator (Maybe TrendmicroConnectorOperator)

The operation to be performed on the provided Trend Micro source fields.

connectorOperator_veeva :: Lens' ConnectorOperator (Maybe VeevaConnectorOperator)

The operation to be performed on the provided Veeva source fields.