Network.AWS.AppFlow.Types.FieldTypeDetails

Description

Documentation

data FieldTypeDetails

Contains details regarding the supported field type and the operators that can be applied for filtering.

See: newFieldTypeDetails smart constructor.

Constructors

FieldTypeDetails'

Fields

supportedValues :: Maybe [Text]

The list of values that a field can contain. For example, a Boolean fieldType can have two values: "true" and "false".

fieldType :: Text

The type of field, such as string, integer, date, and so on.

filterOperators :: [Operator]

The list of operators supported by a field.

newFieldTypeDetails

Arguments

:: Text
$sel:fieldType:FieldTypeDetails'

-> FieldTypeDetails

Create a value of FieldTypeDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:supportedValues:FieldTypeDetails', fieldTypeDetails_supportedValues - The list of values that a field can contain. For example, a Boolean fieldType can have two values: "true" and "false".

$sel:fieldType:FieldTypeDetails', fieldTypeDetails_fieldType - The type of field, such as string, integer, date, and so on.

$sel:filterOperators:FieldTypeDetails', fieldTypeDetails_filterOperators - The list of operators supported by a field.

fieldTypeDetails_supportedValues :: Lens' FieldTypeDetails (Maybe [Text])

The list of values that a field can contain. For example, a Boolean fieldType can have two values: "true" and "false".

fieldTypeDetails_fieldType :: Lens' FieldTypeDetails Text

The type of field, such as string, integer, date, and so on.

fieldTypeDetails_filterOperators :: Lens' FieldTypeDetails [Operator]

The list of operators supported by a field.