Network.AWS.AppFlow.Types.FlowDefinition

Description

Documentation

data FlowDefinition

The properties of the flow, such as its source, destination, trigger type, and so on.

See: newFlowDefinition smart constructor.

Constructors

FlowDefinition'

Fields

lastUpdatedBy :: Maybe Text

Specifies the account user name that most recently updated the flow.

flowArn :: Maybe Text

The flow's Amazon Resource Name (ARN).

lastUpdatedAt :: Maybe POSIX

Specifies when the flow was last updated.

createdAt :: Maybe POSIX

Specifies when the flow was created.

triggerType :: Maybe TriggerType

Specifies the type of flow trigger. This can be OnDemand, Scheduled, or Event.

createdBy :: Maybe Text

The ARN of the user who created the flow.

sourceConnectorType :: Maybe ConnectorType

Specifies the source connector type, such as Salesforce, Amazon S3, Amplitude, and so on.

destinationConnectorType :: Maybe ConnectorType

Specifies the destination connector type, such as Salesforce, Amazon S3, Amplitude, and so on.

flowName :: Maybe Text

The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

lastRunExecutionDetails :: Maybe ExecutionDetails

Describes the details of the most recent flow run.

flowStatus :: Maybe FlowStatus

Indicates the current status of the flow.

description :: Maybe Text

A user-entered description of the flow.

tags :: Maybe (HashMap Text Text)

The tags used to organize, track, or control access for your flow.

newFlowDefinition :: FlowDefinition

Create a value of FlowDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedBy:FlowDefinition', flowDefinition_lastUpdatedBy - Specifies the account user name that most recently updated the flow.

$sel:flowArn:FlowDefinition', flowDefinition_flowArn - The flow's Amazon Resource Name (ARN).

$sel:lastUpdatedAt:FlowDefinition', flowDefinition_lastUpdatedAt - Specifies when the flow was last updated.

$sel:createdAt:FlowDefinition', flowDefinition_createdAt - Specifies when the flow was created.

$sel:triggerType:FlowDefinition', flowDefinition_triggerType - Specifies the type of flow trigger. This can be OnDemand, Scheduled, or Event.

$sel:createdBy:FlowDefinition', flowDefinition_createdBy - The ARN of the user who created the flow.

$sel:sourceConnectorType:FlowDefinition', flowDefinition_sourceConnectorType - Specifies the source connector type, such as Salesforce, Amazon S3, Amplitude, and so on.

$sel:destinationConnectorType:FlowDefinition', flowDefinition_destinationConnectorType - Specifies the destination connector type, such as Salesforce, Amazon S3, Amplitude, and so on.

$sel:flowName:FlowDefinition', flowDefinition_flowName - The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

$sel:lastRunExecutionDetails:FlowDefinition', flowDefinition_lastRunExecutionDetails - Describes the details of the most recent flow run.

$sel:flowStatus:FlowDefinition', flowDefinition_flowStatus - Indicates the current status of the flow.

$sel:description:FlowDefinition', flowDefinition_description - A user-entered description of the flow.

$sel:tags:FlowDefinition', flowDefinition_tags - The tags used to organize, track, or control access for your flow.

flowDefinition_lastUpdatedBy :: Lens' FlowDefinition (Maybe Text)

Specifies the account user name that most recently updated the flow.

flowDefinition_flowArn :: Lens' FlowDefinition (Maybe Text)

The flow's Amazon Resource Name (ARN).

flowDefinition_lastUpdatedAt :: Lens' FlowDefinition (Maybe UTCTime)

Specifies when the flow was last updated.

flowDefinition_createdAt :: Lens' FlowDefinition (Maybe UTCTime)

Specifies when the flow was created.

flowDefinition_triggerType :: Lens' FlowDefinition (Maybe TriggerType)

Specifies the type of flow trigger. This can be OnDemand, Scheduled, or Event.

flowDefinition_createdBy :: Lens' FlowDefinition (Maybe Text)

The ARN of the user who created the flow.

flowDefinition_sourceConnectorType :: Lens' FlowDefinition (Maybe ConnectorType)

Specifies the source connector type, such as Salesforce, Amazon S3, Amplitude, and so on.

flowDefinition_destinationConnectorType :: Lens' FlowDefinition (Maybe ConnectorType)

Specifies the destination connector type, such as Salesforce, Amazon S3, Amplitude, and so on.

flowDefinition_flowName :: Lens' FlowDefinition (Maybe Text)

The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

flowDefinition_lastRunExecutionDetails :: Lens' FlowDefinition (Maybe ExecutionDetails)

Describes the details of the most recent flow run.

flowDefinition_flowStatus :: Lens' FlowDefinition (Maybe FlowStatus)

Indicates the current status of the flow.

flowDefinition_description :: Lens' FlowDefinition (Maybe Text)

A user-entered description of the flow.

flowDefinition_tags :: Lens' FlowDefinition (Maybe (HashMap Text Text))

The tags used to organize, track, or control access for your flow.