Network.AWS.AppFlow.Types.GoogleAnalyticsMetadata

Description

Documentation

data GoogleAnalyticsMetadata

The connector metadata specific to Google Analytics.

See: newGoogleAnalyticsMetadata smart constructor.

Constructors

GoogleAnalyticsMetadata'

Fields

oAuthScopes :: Maybe [Text]

The desired authorization scope for the Google Analytics account.

newGoogleAnalyticsMetadata :: GoogleAnalyticsMetadata

Create a value of GoogleAnalyticsMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:oAuthScopes:GoogleAnalyticsMetadata', googleAnalyticsMetadata_oAuthScopes - The desired authorization scope for the Google Analytics account.

googleAnalyticsMetadata_oAuthScopes :: Lens' GoogleAnalyticsMetadata (Maybe [Text])

The desired authorization scope for the Google Analytics account.