Network.AWS.AppFlow.Types.OAuthCredentials

Description

Documentation

data OAuthCredentials

The OAuth credentials required for OAuth type authentication.

See: newOAuthCredentials smart constructor.

Constructors

OAuthCredentials'

Fields

accessToken :: Maybe (Sensitive Text)

The access token used to access protected SAPOData resources.

refreshToken :: Maybe Text

The refresh token used to refresh expired access token.

oAuthRequest :: Maybe ConnectorOAuthRequest

The OAuth requirement needed to request security tokens from the connector endpoint.

clientId :: Text

The identifier for the desired client.

clientSecret :: Sensitive Text

The client secret used by the OAuth client to authenticate to the authorization server.

newOAuthCredentials

Arguments

:: Text
$sel:clientId:OAuthCredentials'

-> Text
$sel:clientSecret:OAuthCredentials'

-> OAuthCredentials

Create a value of OAuthCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessToken:OAuthCredentials', oAuthCredentials_accessToken - The access token used to access protected SAPOData resources.

$sel:refreshToken:OAuthCredentials', oAuthCredentials_refreshToken - The refresh token used to refresh expired access token.

$sel:oAuthRequest:OAuthCredentials', oAuthCredentials_oAuthRequest - The OAuth requirement needed to request security tokens from the connector endpoint.

$sel:clientId:OAuthCredentials', oAuthCredentials_clientId - The identifier for the desired client.

$sel:clientSecret:OAuthCredentials', oAuthCredentials_clientSecret - The client secret used by the OAuth client to authenticate to the authorization server.

oAuthCredentials_accessToken :: Lens' OAuthCredentials (Maybe Text)

The access token used to access protected SAPOData resources.

oAuthCredentials_refreshToken :: Lens' OAuthCredentials (Maybe Text)

The refresh token used to refresh expired access token.

oAuthCredentials_oAuthRequest :: Lens' OAuthCredentials (Maybe ConnectorOAuthRequest)

The OAuth requirement needed to request security tokens from the connector endpoint.

oAuthCredentials_clientId :: Lens' OAuthCredentials Text

The identifier for the desired client.

oAuthCredentials_clientSecret :: Lens' OAuthCredentials Text

The client secret used by the OAuth client to authenticate to the authorization server.