Network.AWS.AppFlow.Types.SnowflakeDestinationProperties

Description

Documentation

data SnowflakeDestinationProperties

The properties that are applied when Snowflake is being used as a destination.

See: newSnowflakeDestinationProperties smart constructor.

Constructors

SnowflakeDestinationProperties'

Fields

bucketPrefix :: Maybe Text

The object key for the destination bucket in which Amazon AppFlow places the files.

errorHandlingConfig :: Maybe ErrorHandlingConfig

The settings that determine how Amazon AppFlow handles an error when placing data in the Snowflake destination. For example, this setting would determine if the flow should fail after one insertion error, or continue and attempt to insert every record regardless of the initial failure. ErrorHandlingConfig is a part of the destination connector details.

object' :: Text

The object specified in the Snowflake flow destination.

intermediateBucketName :: Text

The intermediate bucket that Amazon AppFlow uses when moving data into Snowflake.

newSnowflakeDestinationProperties

Arguments

:: Text
$sel:object':SnowflakeDestinationProperties'

-> Text
$sel:intermediateBucketName:SnowflakeDestinationProperties'

-> SnowflakeDestinationProperties

Create a value of SnowflakeDestinationProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketPrefix:SnowflakeDestinationProperties', snowflakeDestinationProperties_bucketPrefix - The object key for the destination bucket in which Amazon AppFlow places the files.

$sel:errorHandlingConfig:SnowflakeDestinationProperties', snowflakeDestinationProperties_errorHandlingConfig - The settings that determine how Amazon AppFlow handles an error when placing data in the Snowflake destination. For example, this setting would determine if the flow should fail after one insertion error, or continue and attempt to insert every record regardless of the initial failure. ErrorHandlingConfig is a part of the destination connector details.

$sel:object':SnowflakeDestinationProperties', snowflakeDestinationProperties_object - The object specified in the Snowflake flow destination.

$sel:intermediateBucketName:SnowflakeDestinationProperties', snowflakeDestinationProperties_intermediateBucketName - The intermediate bucket that Amazon AppFlow uses when moving data into Snowflake.

snowflakeDestinationProperties_bucketPrefix :: Lens' SnowflakeDestinationProperties (Maybe Text)

The object key for the destination bucket in which Amazon AppFlow places the files.

snowflakeDestinationProperties_errorHandlingConfig :: Lens' SnowflakeDestinationProperties (Maybe ErrorHandlingConfig)

The settings that determine how Amazon AppFlow handles an error when placing data in the Snowflake destination. For example, this setting would determine if the flow should fail after one insertion error, or continue and attempt to insert every record regardless of the initial failure. ErrorHandlingConfig is a part of the destination connector details.

snowflakeDestinationProperties_object :: Lens' SnowflakeDestinationProperties Text

The object specified in the Snowflake flow destination.

snowflakeDestinationProperties_intermediateBucketName :: Lens' SnowflakeDestinationProperties Text

The intermediate bucket that Amazon AppFlow uses when moving data into Snowflake.