Network.AWS.AppFlow.Types.SourceConnectorProperties

Description

Documentation

data SourceConnectorProperties

Specifies the information that is required to query a particular connector.

See: newSourceConnectorProperties smart constructor.

Constructors

SourceConnectorProperties'

Fields

serviceNow :: Maybe ServiceNowSourceProperties

Specifies the information that is required for querying ServiceNow.

dynatrace :: Maybe DynatraceSourceProperties

Specifies the information that is required for querying Dynatrace.

marketo :: Maybe MarketoSourceProperties

Specifies the information that is required for querying Marketo.

slack :: Maybe SlackSourceProperties

Specifies the information that is required for querying Slack.

singular :: Maybe SingularSourceProperties

Specifies the information that is required for querying Singular.

inforNexus :: Maybe InforNexusSourceProperties

Specifies the information that is required for querying Infor Nexus.

amplitude :: Maybe AmplitudeSourceProperties

Specifies the information that is required for querying Amplitude.

datadog :: Maybe DatadogSourceProperties

Specifies the information that is required for querying Datadog.

googleAnalytics :: Maybe GoogleAnalyticsSourceProperties

Specifies the information that is required for querying Google Analytics.

sAPOData :: Maybe SAPODataSourceProperties

salesforce :: Maybe SalesforceSourceProperties

Specifies the information that is required for querying Salesforce.

zendesk :: Maybe ZendeskSourceProperties

Specifies the information that is required for querying Zendesk.

s3 :: Maybe S3SourceProperties

Specifies the information that is required for querying Amazon S3.

trendmicro :: Maybe TrendmicroSourceProperties

Specifies the information that is required for querying Trend Micro.

veeva :: Maybe VeevaSourceProperties

Specifies the information that is required for querying Veeva.

newSourceConnectorProperties :: SourceConnectorProperties

Create a value of SourceConnectorProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceNow:SourceConnectorProperties', sourceConnectorProperties_serviceNow - Specifies the information that is required for querying ServiceNow.

$sel:dynatrace:SourceConnectorProperties', sourceConnectorProperties_dynatrace - Specifies the information that is required for querying Dynatrace.

$sel:marketo:SourceConnectorProperties', sourceConnectorProperties_marketo - Specifies the information that is required for querying Marketo.

$sel:slack:SourceConnectorProperties', sourceConnectorProperties_slack - Specifies the information that is required for querying Slack.

$sel:singular:SourceConnectorProperties', sourceConnectorProperties_singular - Specifies the information that is required for querying Singular.

$sel:inforNexus:SourceConnectorProperties', sourceConnectorProperties_inforNexus - Specifies the information that is required for querying Infor Nexus.

$sel:amplitude:SourceConnectorProperties', sourceConnectorProperties_amplitude - Specifies the information that is required for querying Amplitude.

$sel:datadog:SourceConnectorProperties', sourceConnectorProperties_datadog - Specifies the information that is required for querying Datadog.

$sel:googleAnalytics:SourceConnectorProperties', sourceConnectorProperties_googleAnalytics - Specifies the information that is required for querying Google Analytics.

$sel:sAPOData:SourceConnectorProperties', sourceConnectorProperties_sAPOData - Undocumented member.

$sel:salesforce:SourceConnectorProperties', sourceConnectorProperties_salesforce - Specifies the information that is required for querying Salesforce.

$sel:zendesk:SourceConnectorProperties', sourceConnectorProperties_zendesk - Specifies the information that is required for querying Zendesk.

$sel:s3:SourceConnectorProperties', sourceConnectorProperties_s3 - Specifies the information that is required for querying Amazon S3.

$sel:trendmicro:SourceConnectorProperties', sourceConnectorProperties_trendmicro - Specifies the information that is required for querying Trend Micro.

$sel:veeva:SourceConnectorProperties', sourceConnectorProperties_veeva - Specifies the information that is required for querying Veeva.

sourceConnectorProperties_serviceNow :: Lens' SourceConnectorProperties (Maybe ServiceNowSourceProperties)

Specifies the information that is required for querying ServiceNow.

sourceConnectorProperties_dynatrace :: Lens' SourceConnectorProperties (Maybe DynatraceSourceProperties)

Specifies the information that is required for querying Dynatrace.

sourceConnectorProperties_marketo :: Lens' SourceConnectorProperties (Maybe MarketoSourceProperties)

Specifies the information that is required for querying Marketo.

sourceConnectorProperties_slack :: Lens' SourceConnectorProperties (Maybe SlackSourceProperties)

Specifies the information that is required for querying Slack.

sourceConnectorProperties_singular :: Lens' SourceConnectorProperties (Maybe SingularSourceProperties)

Specifies the information that is required for querying Singular.

sourceConnectorProperties_inforNexus :: Lens' SourceConnectorProperties (Maybe InforNexusSourceProperties)

Specifies the information that is required for querying Infor Nexus.

sourceConnectorProperties_amplitude :: Lens' SourceConnectorProperties (Maybe AmplitudeSourceProperties)

Specifies the information that is required for querying Amplitude.

sourceConnectorProperties_datadog :: Lens' SourceConnectorProperties (Maybe DatadogSourceProperties)

Specifies the information that is required for querying Datadog.

sourceConnectorProperties_googleAnalytics :: Lens' SourceConnectorProperties (Maybe GoogleAnalyticsSourceProperties)

Specifies the information that is required for querying Google Analytics.

sourceConnectorProperties_sAPOData :: Lens' SourceConnectorProperties (Maybe SAPODataSourceProperties)

Undocumented member.

sourceConnectorProperties_salesforce :: Lens' SourceConnectorProperties (Maybe SalesforceSourceProperties)

Specifies the information that is required for querying Salesforce.

sourceConnectorProperties_zendesk :: Lens' SourceConnectorProperties (Maybe ZendeskSourceProperties)

Specifies the information that is required for querying Zendesk.

sourceConnectorProperties_s3 :: Lens' SourceConnectorProperties (Maybe S3SourceProperties)

Specifies the information that is required for querying Amazon S3.

sourceConnectorProperties_trendmicro :: Lens' SourceConnectorProperties (Maybe TrendmicroSourceProperties)

Specifies the information that is required for querying Trend Micro.

sourceConnectorProperties_veeva :: Lens' SourceConnectorProperties (Maybe VeevaSourceProperties)

Specifies the information that is required for querying Veeva.