Network.AWS.AppFlow.Types.VeevaSourceProperties

Description

Documentation

data VeevaSourceProperties

The properties that are applied when using Veeva as a flow source.

See: newVeevaSourceProperties smart constructor.

Constructors

VeevaSourceProperties'

Fields

includeAllVersions :: Maybe Bool

Boolean value to include All Versions of files in Veeva document extract flow.

documentType :: Maybe Text

The document type specified in the Veeva document extract flow.

includeRenditions :: Maybe Bool

Boolean value to include file renditions in Veeva document extract flow.

includeSourceFiles :: Maybe Bool

Boolean value to include source files in Veeva document extract flow.

object' :: Text

The object specified in the Veeva flow source.

newVeevaSourceProperties

Arguments

:: Text
$sel:object':VeevaSourceProperties'

-> VeevaSourceProperties

Create a value of VeevaSourceProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeAllVersions:VeevaSourceProperties', veevaSourceProperties_includeAllVersions - Boolean value to include All Versions of files in Veeva document extract flow.

$sel:documentType:VeevaSourceProperties', veevaSourceProperties_documentType - The document type specified in the Veeva document extract flow.

$sel:includeRenditions:VeevaSourceProperties', veevaSourceProperties_includeRenditions - Boolean value to include file renditions in Veeva document extract flow.

$sel:includeSourceFiles:VeevaSourceProperties', veevaSourceProperties_includeSourceFiles - Boolean value to include source files in Veeva document extract flow.

$sel:object':VeevaSourceProperties', veevaSourceProperties_object - The object specified in the Veeva flow source.

veevaSourceProperties_includeAllVersions :: Lens' VeevaSourceProperties (Maybe Bool)

Boolean value to include All Versions of files in Veeva document extract flow.

veevaSourceProperties_documentType :: Lens' VeevaSourceProperties (Maybe Text)

The document type specified in the Veeva document extract flow.

veevaSourceProperties_includeRenditions :: Lens' VeevaSourceProperties (Maybe Bool)

Boolean value to include file renditions in Veeva document extract flow.

veevaSourceProperties_includeSourceFiles :: Lens' VeevaSourceProperties (Maybe Bool)

Boolean value to include source files in Veeva document extract flow.

veevaSourceProperties_object :: Lens' VeevaSourceProperties Text

The object specified in the Veeva flow source.