Network.AWS.AppFlow.UpdateFlow

Description

Updates an existing flow.

Creating a Request

data UpdateFlow

See: newUpdateFlow smart constructor.

Constructors

UpdateFlow'

Fields

description :: Maybe Text

A description of the flow.

flowName :: Text

The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

triggerConfig :: TriggerConfig

The trigger settings that determine how and when the flow runs.

sourceFlowConfig :: SourceFlowConfig

destinationFlowConfigList :: [DestinationFlowConfig]

The configuration that controls how Amazon AppFlow transfers data to the destination connector.

tasks :: [Task]

A list of tasks that Amazon AppFlow performs while transferring the data in the flow run.

newUpdateFlow

Arguments

:: Text
$sel:flowName:UpdateFlow'

-> TriggerConfig
$sel:triggerConfig:UpdateFlow'

-> SourceFlowConfig
$sel:sourceFlowConfig:UpdateFlow'

-> UpdateFlow

Create a value of UpdateFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:UpdateFlow', updateFlow_description - A description of the flow.

$sel:flowName:UpdateFlow', updateFlow_flowName - The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

$sel:triggerConfig:UpdateFlow', updateFlow_triggerConfig - The trigger settings that determine how and when the flow runs.

$sel:sourceFlowConfig:UpdateFlow', updateFlow_sourceFlowConfig - Undocumented member.

$sel:destinationFlowConfigList:UpdateFlow', updateFlow_destinationFlowConfigList - The configuration that controls how Amazon AppFlow transfers data to the destination connector.

$sel:tasks:UpdateFlow', updateFlow_tasks - A list of tasks that Amazon AppFlow performs while transferring the data in the flow run.

Request Lenses

updateFlow_description :: Lens' UpdateFlow (Maybe Text)

A description of the flow.

updateFlow_flowName :: Lens' UpdateFlow Text

The specified name of the flow. Spaces are not allowed. Use underscores (_) or hyphens (-) only.

updateFlow_triggerConfig :: Lens' UpdateFlow TriggerConfig

The trigger settings that determine how and when the flow runs.

updateFlow_sourceFlowConfig :: Lens' UpdateFlow SourceFlowConfig

Undocumented member.

updateFlow_destinationFlowConfigList :: Lens' UpdateFlow [DestinationFlowConfig]

The configuration that controls how Amazon AppFlow transfers data to the destination connector.

updateFlow_tasks :: Lens' UpdateFlow [Task]

A list of tasks that Amazon AppFlow performs while transferring the data in the flow run.

Destructuring the Response

data UpdateFlowResponse

See: newUpdateFlowResponse smart constructor.

Constructors

UpdateFlowResponse'

Fields

flowStatus :: Maybe FlowStatus

Indicates the current status of the flow.

httpStatus :: Int

The response's http status code.

newUpdateFlowResponse

Arguments

:: Int
$sel:httpStatus:UpdateFlowResponse'

-> UpdateFlowResponse

Create a value of UpdateFlowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowStatus:UpdateFlowResponse', updateFlowResponse_flowStatus - Indicates the current status of the flow.

$sel:httpStatus:UpdateFlowResponse', updateFlowResponse_httpStatus - The response's http status code.

Response Lenses

updateFlowResponse_flowStatus :: Lens' UpdateFlowResponse (Maybe FlowStatus)

Indicates the current status of the flow.

updateFlowResponse_httpStatus :: Lens' UpdateFlowResponse Int

The response's http status code.