Network.AWS.AppIntegrationS.CreateDataIntegration

Description

Creates and persists a DataIntegration resource.

You cannot create a DataIntegration association for a DataIntegration that has been previously associated. Use a different DataIntegration, or recreate the DataIntegration using the CreateDataIntegration API.

Creating a Request

data CreateDataIntegration

See: newCreateDataIntegration smart constructor.

Constructors

CreateDataIntegration'

Fields

scheduleConfig :: Maybe ScheduleConfiguration

The name of the data and how often it should be pulled from the source.

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

kmsKey :: Maybe Text

The KMS key for the DataIntegration.

sourceURI :: Maybe Text

The URI of the data source.

description :: Maybe Text

A description of the DataIntegration.

tags :: Maybe (HashMap Text Text)

One or more tags.

name :: Text

The name of the DataIntegration.

newCreateDataIntegration

Arguments

:: Text
$sel:name:CreateDataIntegration'

-> CreateDataIntegration

Create a value of CreateDataIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduleConfig:CreateDataIntegration', createDataIntegration_scheduleConfig - The name of the data and how often it should be pulled from the source.

$sel:clientToken:CreateDataIntegration', createDataIntegration_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:kmsKey:CreateDataIntegration', createDataIntegration_kmsKey - The KMS key for the DataIntegration.

$sel:sourceURI:CreateDataIntegration', createDataIntegration_sourceURI - The URI of the data source.

$sel:description:CreateDataIntegration', createDataIntegration_description - A description of the DataIntegration.

$sel:tags:CreateDataIntegration', createDataIntegration_tags - One or more tags.

$sel:name:CreateDataIntegration', createDataIntegration_name - The name of the DataIntegration.

Request Lenses

createDataIntegration_scheduleConfig :: Lens' CreateDataIntegration (Maybe ScheduleConfiguration)

The name of the data and how often it should be pulled from the source.

createDataIntegration_clientToken :: Lens' CreateDataIntegration (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createDataIntegration_kmsKey :: Lens' CreateDataIntegration (Maybe Text)

The KMS key for the DataIntegration.

createDataIntegration_sourceURI :: Lens' CreateDataIntegration (Maybe Text)

The URI of the data source.

createDataIntegration_description :: Lens' CreateDataIntegration (Maybe Text)

A description of the DataIntegration.

createDataIntegration_tags :: Lens' CreateDataIntegration (Maybe (HashMap Text Text))

One or more tags.

createDataIntegration_name :: Lens' CreateDataIntegration Text

The name of the DataIntegration.

Destructuring the Response

data CreateDataIntegrationResponse

See: newCreateDataIntegrationResponse smart constructor.

Constructors

CreateDataIntegrationResponse'

Fields

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

kmsKey :: Maybe Text

The KMS key for the DataIntegration.

arn :: Maybe Text

The Amazon Resource Name (ARN)

scheduleConfiguration :: Maybe ScheduleConfiguration

The name of the data and how often it should be pulled from the source.

name :: Maybe Text

The name of the DataIntegration.

id :: Maybe Text

A unique identifier.

sourceURI :: Maybe Text

The URI of the data source.

description :: Maybe Text

A description of the DataIntegration.

tags :: Maybe (HashMap Text Text)

One or more tags.

httpStatus :: Int

The response's http status code.

newCreateDataIntegrationResponse

Arguments

:: Int
$sel:httpStatus:CreateDataIntegrationResponse'

-> CreateDataIntegrationResponse

Create a value of CreateDataIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateDataIntegration', createDataIntegrationResponse_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:kmsKey:CreateDataIntegration', createDataIntegrationResponse_kmsKey - The KMS key for the DataIntegration.

$sel:arn:CreateDataIntegrationResponse', createDataIntegrationResponse_arn - The Amazon Resource Name (ARN)

$sel:scheduleConfiguration:CreateDataIntegrationResponse', createDataIntegrationResponse_scheduleConfiguration - The name of the data and how often it should be pulled from the source.

$sel:name:CreateDataIntegration', createDataIntegrationResponse_name - The name of the DataIntegration.

$sel:id:CreateDataIntegrationResponse', createDataIntegrationResponse_id - A unique identifier.

$sel:sourceURI:CreateDataIntegration', createDataIntegrationResponse_sourceURI - The URI of the data source.

$sel:description:CreateDataIntegration', createDataIntegrationResponse_description - A description of the DataIntegration.

$sel:tags:CreateDataIntegration', createDataIntegrationResponse_tags - One or more tags.

$sel:httpStatus:CreateDataIntegrationResponse', createDataIntegrationResponse_httpStatus - The response's http status code.

Response Lenses

createDataIntegrationResponse_clientToken :: Lens' CreateDataIntegrationResponse (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createDataIntegrationResponse_kmsKey :: Lens' CreateDataIntegrationResponse (Maybe Text)

The KMS key for the DataIntegration.

createDataIntegrationResponse_arn :: Lens' CreateDataIntegrationResponse (Maybe Text)

The Amazon Resource Name (ARN)

createDataIntegrationResponse_scheduleConfiguration :: Lens' CreateDataIntegrationResponse (Maybe ScheduleConfiguration)

The name of the data and how often it should be pulled from the source.

createDataIntegrationResponse_name :: Lens' CreateDataIntegrationResponse (Maybe Text)

The name of the DataIntegration.

createDataIntegrationResponse_id :: Lens' CreateDataIntegrationResponse (Maybe Text)

A unique identifier.

createDataIntegrationResponse_sourceURI :: Lens' CreateDataIntegrationResponse (Maybe Text)

The URI of the data source.

createDataIntegrationResponse_description :: Lens' CreateDataIntegrationResponse (Maybe Text)

A description of the DataIntegration.

createDataIntegrationResponse_tags :: Lens' CreateDataIntegrationResponse (Maybe (HashMap Text Text))

One or more tags.

createDataIntegrationResponse_httpStatus :: Lens' CreateDataIntegrationResponse Int

The response's http status code.