Network.AWS.AppIntegrationS.DeleteDataIntegration

Description

Deletes the DataIntegration. Only DataIntegrations that don't have any DataIntegrationAssociations can be deleted. Deleting a DataIntegration also deletes the underlying Amazon AppFlow flow and service linked role.

You cannot create a DataIntegration association for a DataIntegration that has been previously associated. Use a different DataIntegration, or recreate the DataIntegration using the CreateDataIntegration API.

Creating a Request

data DeleteDataIntegration

See: newDeleteDataIntegration smart constructor.

Constructors

DeleteDataIntegration'

Fields

dataIntegrationIdentifier :: Text

A unique identifier for the DataIntegration.

newDeleteDataIntegration

Arguments

:: Text
$sel:dataIntegrationIdentifier:DeleteDataIntegration'

-> DeleteDataIntegration

Create a value of DeleteDataIntegration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataIntegrationIdentifier:DeleteDataIntegration', deleteDataIntegration_dataIntegrationIdentifier - A unique identifier for the DataIntegration.

Request Lenses

deleteDataIntegration_dataIntegrationIdentifier :: Lens' DeleteDataIntegration Text

A unique identifier for the DataIntegration.

Destructuring the Response

data DeleteDataIntegrationResponse

See: newDeleteDataIntegrationResponse smart constructor.

Constructors

DeleteDataIntegrationResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteDataIntegrationResponse

Arguments

:: Int
$sel:httpStatus:DeleteDataIntegrationResponse'

-> DeleteDataIntegrationResponse

Create a value of DeleteDataIntegrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteDataIntegrationResponse', deleteDataIntegrationResponse_httpStatus - The response's http status code.

Response Lenses

deleteDataIntegrationResponse_httpStatus :: Lens' DeleteDataIntegrationResponse Int

The response's http status code.