Network.AWS.AppIntegrationS.Lens

Description

Operations

GetEventIntegration

getEventIntegration_name :: Lens' GetEventIntegration Text

The name of the event integration.

getEventIntegrationResponse_eventBridgeBus :: Lens' GetEventIntegrationResponse (Maybe Text)

The EventBridge bus.

getEventIntegrationResponse_eventFilter :: Lens' GetEventIntegrationResponse (Maybe EventFilter)

The event filter.

getEventIntegrationResponse_eventIntegrationArn :: Lens' GetEventIntegrationResponse (Maybe Text)

The Amazon Resource Name (ARN) for the event integration.

getEventIntegrationResponse_name :: Lens' GetEventIntegrationResponse (Maybe Text)

The name of the event integration.

getEventIntegrationResponse_description :: Lens' GetEventIntegrationResponse (Maybe Text)

The description of the event integration.

getEventIntegrationResponse_tags :: Lens' GetEventIntegrationResponse (Maybe (HashMap Text Text))

One or more tags.

getEventIntegrationResponse_httpStatus :: Lens' GetEventIntegrationResponse Int

The response's http status code.

ListDataIntegrations

listDataIntegrations_nextToken :: Lens' ListDataIntegrations (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listDataIntegrations_maxResults :: Lens' ListDataIntegrations (Maybe Natural)

The maximum number of results to return per page.

listDataIntegrationsResponse_dataIntegrations :: Lens' ListDataIntegrationsResponse (Maybe (NonEmpty DataIntegrationSummary))

The DataIntegrations associated with this account.

listDataIntegrationsResponse_nextToken :: Lens' ListDataIntegrationsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listDataIntegrationsResponse_httpStatus :: Lens' ListDataIntegrationsResponse Int

The response's http status code.

CreateDataIntegration

createDataIntegration_scheduleConfig :: Lens' CreateDataIntegration (Maybe ScheduleConfiguration)

The name of the data and how often it should be pulled from the source.

createDataIntegration_clientToken :: Lens' CreateDataIntegration (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createDataIntegration_kmsKey :: Lens' CreateDataIntegration (Maybe Text)

The KMS key for the DataIntegration.

createDataIntegration_sourceURI :: Lens' CreateDataIntegration (Maybe Text)

The URI of the data source.

createDataIntegration_description :: Lens' CreateDataIntegration (Maybe Text)

A description of the DataIntegration.

createDataIntegration_tags :: Lens' CreateDataIntegration (Maybe (HashMap Text Text))

One or more tags.

createDataIntegration_name :: Lens' CreateDataIntegration Text

The name of the DataIntegration.

createDataIntegrationResponse_clientToken :: Lens' CreateDataIntegrationResponse (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createDataIntegrationResponse_kmsKey :: Lens' CreateDataIntegrationResponse (Maybe Text)

The KMS key for the DataIntegration.

createDataIntegrationResponse_arn :: Lens' CreateDataIntegrationResponse (Maybe Text)

The Amazon Resource Name (ARN)

createDataIntegrationResponse_scheduleConfiguration :: Lens' CreateDataIntegrationResponse (Maybe ScheduleConfiguration)

The name of the data and how often it should be pulled from the source.

createDataIntegrationResponse_name :: Lens' CreateDataIntegrationResponse (Maybe Text)

The name of the DataIntegration.

createDataIntegrationResponse_id :: Lens' CreateDataIntegrationResponse (Maybe Text)

A unique identifier.

createDataIntegrationResponse_sourceURI :: Lens' CreateDataIntegrationResponse (Maybe Text)

The URI of the data source.

createDataIntegrationResponse_description :: Lens' CreateDataIntegrationResponse (Maybe Text)

A description of the DataIntegration.

createDataIntegrationResponse_tags :: Lens' CreateDataIntegrationResponse (Maybe (HashMap Text Text))

One or more tags.

createDataIntegrationResponse_httpStatus :: Lens' CreateDataIntegrationResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

Information about the tags.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

ListEventIntegrationAssociations

listEventIntegrationAssociations_nextToken :: Lens' ListEventIntegrationAssociations (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listEventIntegrationAssociations_maxResults :: Lens' ListEventIntegrationAssociations (Maybe Natural)

The maximum number of results to return per page.

listEventIntegrationAssociations_eventIntegrationName :: Lens' ListEventIntegrationAssociations Text

The name of the event integration.

listEventIntegrationAssociationsResponse_eventIntegrationAssociations :: Lens' ListEventIntegrationAssociationsResponse (Maybe (NonEmpty EventIntegrationAssociation))

The event integration associations.

listEventIntegrationAssociationsResponse_nextToken :: Lens' ListEventIntegrationAssociationsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listEventIntegrationAssociationsResponse_httpStatus :: Lens' ListEventIntegrationAssociationsResponse Int

The response's http status code.

GetDataIntegration

getDataIntegration_identifier :: Lens' GetDataIntegration Text

A unique identifier.

getDataIntegrationResponse_kmsKey :: Lens' GetDataIntegrationResponse (Maybe Text)

The KMS key for the DataIntegration.

getDataIntegrationResponse_arn :: Lens' GetDataIntegrationResponse (Maybe Text)

The Amazon Resource Name (ARN) for the DataIntegration.

getDataIntegrationResponse_scheduleConfiguration :: Lens' GetDataIntegrationResponse (Maybe ScheduleConfiguration)

The name of the data and how often it should be pulled from the source.

getDataIntegrationResponse_name :: Lens' GetDataIntegrationResponse (Maybe Text)

The name of the DataIntegration.

getDataIntegrationResponse_id :: Lens' GetDataIntegrationResponse (Maybe Text)

A unique identifier.

getDataIntegrationResponse_sourceURI :: Lens' GetDataIntegrationResponse (Maybe Text)

The URI of the data source.

getDataIntegrationResponse_description :: Lens' GetDataIntegrationResponse (Maybe Text)

The KMS key for the DataIntegration.

getDataIntegrationResponse_tags :: Lens' GetDataIntegrationResponse (Maybe (HashMap Text Text))

One or more tags.

getDataIntegrationResponse_httpStatus :: Lens' GetDataIntegrationResponse Int

The response's http status code.

ListEventIntegrations

listEventIntegrations_nextToken :: Lens' ListEventIntegrations (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listEventIntegrations_maxResults :: Lens' ListEventIntegrations (Maybe Natural)

The maximum number of results to return per page.

listEventIntegrationsResponse_eventIntegrations :: Lens' ListEventIntegrationsResponse (Maybe (NonEmpty EventIntegration))

The event integrations.

listEventIntegrationsResponse_nextToken :: Lens' ListEventIntegrationsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listEventIntegrationsResponse_httpStatus :: Lens' ListEventIntegrationsResponse Int

The response's http status code.

DeleteEventIntegration

deleteEventIntegration_name :: Lens' DeleteEventIntegration Text

The name of the event integration.

deleteEventIntegrationResponse_httpStatus :: Lens' DeleteEventIntegrationResponse Int

The response's http status code.

UpdateEventIntegration

updateEventIntegration_description :: Lens' UpdateEventIntegration (Maybe Text)

The description of the event inegration.

updateEventIntegration_name :: Lens' UpdateEventIntegration Text

The name of the event integration.

updateEventIntegrationResponse_httpStatus :: Lens' UpdateEventIntegrationResponse Int

The response's http status code.

DeleteDataIntegration

deleteDataIntegration_dataIntegrationIdentifier :: Lens' DeleteDataIntegration Text

A unique identifier for the DataIntegration.

deleteDataIntegrationResponse_httpStatus :: Lens' DeleteDataIntegrationResponse Int

The response's http status code.

UpdateDataIntegration

updateDataIntegration_name :: Lens' UpdateDataIntegration (Maybe Text)

The name of the DataIntegration.

updateDataIntegration_description :: Lens' UpdateDataIntegration (Maybe Text)

A description of the DataIntegration.

updateDataIntegration_identifier :: Lens' UpdateDataIntegration Text

A unique identifier for the DataIntegration.

updateDataIntegrationResponse_httpStatus :: Lens' UpdateDataIntegrationResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

One or more tags.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The tag keys.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

ListDataIntegrationAssociations

listDataIntegrationAssociations_nextToken :: Lens' ListDataIntegrationAssociations (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listDataIntegrationAssociations_maxResults :: Lens' ListDataIntegrationAssociations (Maybe Natural)

The maximum number of results to return per page.

listDataIntegrationAssociations_dataIntegrationIdentifier :: Lens' ListDataIntegrationAssociations Text

A unique identifier for the DataIntegration.

listDataIntegrationAssociationsResponse_dataIntegrationAssociations :: Lens' ListDataIntegrationAssociationsResponse (Maybe (NonEmpty DataIntegrationAssociationSummary))

The Amazon Resource Name (ARN) and unique ID of the DataIntegration association.

listDataIntegrationAssociationsResponse_nextToken :: Lens' ListDataIntegrationAssociationsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listDataIntegrationAssociationsResponse_httpStatus :: Lens' ListDataIntegrationAssociationsResponse Int

The response's http status code.

CreateEventIntegration

createEventIntegration_clientToken :: Lens' CreateEventIntegration (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

createEventIntegration_description :: Lens' CreateEventIntegration (Maybe Text)

The description of the event integration.

createEventIntegration_tags :: Lens' CreateEventIntegration (Maybe (HashMap Text Text))

One or more tags.

createEventIntegration_name :: Lens' CreateEventIntegration Text

The name of the event integration.

createEventIntegration_eventFilter :: Lens' CreateEventIntegration EventFilter

The event filter.

createEventIntegration_eventBridgeBus :: Lens' CreateEventIntegration Text

The EventBridge bus.

createEventIntegrationResponse_eventIntegrationArn :: Lens' CreateEventIntegrationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the event integration.

createEventIntegrationResponse_httpStatus :: Lens' CreateEventIntegrationResponse Int

The response's http status code.

Types

DataIntegrationAssociationSummary

dataIntegrationAssociationSummary_clientId :: Lens' DataIntegrationAssociationSummary (Maybe Text)

The identifier for teh client that is associated with the DataIntegration association.

dataIntegrationAssociationSummary_dataIntegrationAssociationArn :: Lens' DataIntegrationAssociationSummary (Maybe Text)

The Amazon Resource Name (ARN) of the DataIntegration association.

dataIntegrationAssociationSummary_dataIntegrationArn :: Lens' DataIntegrationAssociationSummary (Maybe Text)

The Amazon Resource Name (ARN)of the DataIntegration.

DataIntegrationSummary

dataIntegrationSummary_arn :: Lens' DataIntegrationSummary (Maybe Text)

The Amazon Resource Name (ARN) of the DataIntegration.

dataIntegrationSummary_name :: Lens' DataIntegrationSummary (Maybe Text)

The name of the DataIntegration.

dataIntegrationSummary_sourceURI :: Lens' DataIntegrationSummary (Maybe Text)

The URI of the data source.

EventFilter

eventFilter_source :: Lens' EventFilter Text

The source of the events.

EventIntegration

eventIntegration_eventBridgeBus :: Lens' EventIntegration (Maybe Text)

The Amazon EventBridge bus for the event integration.

eventIntegration_eventFilter :: Lens' EventIntegration (Maybe EventFilter)

The event integration filter.

eventIntegration_eventIntegrationArn :: Lens' EventIntegration (Maybe Text)

The Amazon Resource Name (ARN) of the event integration.

eventIntegration_name :: Lens' EventIntegration (Maybe Text)

The name of the event integration.

eventIntegration_description :: Lens' EventIntegration (Maybe Text)

The event integration description.

eventIntegration_tags :: Lens' EventIntegration (Maybe (HashMap Text Text))

The tags.

EventIntegrationAssociation

eventIntegrationAssociation_clientId :: Lens' EventIntegrationAssociation (Maybe Text)

The identifier for the client that is associated with the event integration.

eventIntegrationAssociation_eventIntegrationName :: Lens' EventIntegrationAssociation (Maybe Text)

The name of the event integration.

eventIntegrationAssociation_clientAssociationMetadata :: Lens' EventIntegrationAssociation (Maybe (HashMap Text Text))

The metadata associated with the client.

eventIntegrationAssociation_eventIntegrationAssociationId :: Lens' EventIntegrationAssociation (Maybe Text)

The identifier for the event integration association.

eventIntegrationAssociation_eventIntegrationAssociationArn :: Lens' EventIntegrationAssociation (Maybe Text)

The Amazon Resource Name (ARN) for the event integration association.

eventIntegrationAssociation_eventBridgeRuleName :: Lens' EventIntegrationAssociation (Maybe Text)

The name of the EventBridge rule.

ScheduleConfiguration

scheduleConfiguration_scheduleExpression :: Lens' ScheduleConfiguration (Maybe Text)

How often the data should be pulled from data source.

scheduleConfiguration_object :: Lens' ScheduleConfiguration (Maybe Text)

The name of the object to pull from the data source.

scheduleConfiguration_firstExecutionFrom :: Lens' ScheduleConfiguration (Maybe Text)

The start date for objects to import in the first flow run.