Network.AWS.AppIntegrationS.ListDataIntegrationAssociations

Description

Returns a paginated list of DataIntegration associations in the account.

You cannot create a DataIntegration association for a DataIntegration that has been previously associated. Use a different DataIntegration, or recreate the DataIntegration using the CreateDataIntegration API.

Creating a Request

data ListDataIntegrationAssociations

See: newListDataIntegrationAssociations smart constructor.

Constructors

ListDataIntegrationAssociations'

Fields

nextToken :: Maybe Text

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per page.

dataIntegrationIdentifier :: Text

A unique identifier for the DataIntegration.

newListDataIntegrationAssociations

Arguments

:: Text
$sel:dataIntegrationIdentifier:ListDataIntegrationAssociations'

-> ListDataIntegrationAssociations

Create a value of ListDataIntegrationAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDataIntegrationAssociations', listDataIntegrationAssociations_nextToken - The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

$sel:maxResults:ListDataIntegrationAssociations', listDataIntegrationAssociations_maxResults - The maximum number of results to return per page.

$sel:dataIntegrationIdentifier:ListDataIntegrationAssociations', listDataIntegrationAssociations_dataIntegrationIdentifier - A unique identifier for the DataIntegration.

Request Lenses

listDataIntegrationAssociations_nextToken :: Lens' ListDataIntegrationAssociations (Maybe Text)

The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

listDataIntegrationAssociations_maxResults :: Lens' ListDataIntegrationAssociations (Maybe Natural)

The maximum number of results to return per page.

listDataIntegrationAssociations_dataIntegrationIdentifier :: Lens' ListDataIntegrationAssociations Text

A unique identifier for the DataIntegration.

Destructuring the Response

data ListDataIntegrationAssociationsResponse

See: newListDataIntegrationAssociationsResponse smart constructor.

Constructors

ListDataIntegrationAssociationsResponse'

Fields

dataIntegrationAssociations :: Maybe (NonEmpty DataIntegrationAssociationSummary)

The Amazon Resource Name (ARN) and unique ID of the DataIntegration association.

nextToken :: Maybe Text

If there are additional results, this is the token for the next set of results.

httpStatus :: Int

The response's http status code.

newListDataIntegrationAssociationsResponse

Arguments

:: Int
$sel:httpStatus:ListDataIntegrationAssociationsResponse'

-> ListDataIntegrationAssociationsResponse

Create a value of ListDataIntegrationAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataIntegrationAssociations:ListDataIntegrationAssociationsResponse', listDataIntegrationAssociationsResponse_dataIntegrationAssociations - The Amazon Resource Name (ARN) and unique ID of the DataIntegration association.

$sel:nextToken:ListDataIntegrationAssociations', listDataIntegrationAssociationsResponse_nextToken - If there are additional results, this is the token for the next set of results.

$sel:httpStatus:ListDataIntegrationAssociationsResponse', listDataIntegrationAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

listDataIntegrationAssociationsResponse_dataIntegrationAssociations :: Lens' ListDataIntegrationAssociationsResponse (Maybe (NonEmpty DataIntegrationAssociationSummary))

The Amazon Resource Name (ARN) and unique ID of the DataIntegration association.

listDataIntegrationAssociationsResponse_nextToken :: Lens' ListDataIntegrationAssociationsResponse (Maybe Text)

If there are additional results, this is the token for the next set of results.

listDataIntegrationAssociationsResponse_httpStatus :: Lens' ListDataIntegrationAssociationsResponse Int

The response's http status code.