Network.AWS.AppIntegrationS.Types.DataIntegrationAssociationSummary

Description

Documentation

data DataIntegrationAssociationSummary

Summary information about the DataIntegration association.

See: newDataIntegrationAssociationSummary smart constructor.

Constructors

DataIntegrationAssociationSummary'

Fields

clientId :: Maybe Text

The identifier for teh client that is associated with the DataIntegration association.

dataIntegrationAssociationArn :: Maybe Text

The Amazon Resource Name (ARN) of the DataIntegration association.

dataIntegrationArn :: Maybe Text

The Amazon Resource Name (ARN)of the DataIntegration.

newDataIntegrationAssociationSummary :: DataIntegrationAssociationSummary

Create a value of DataIntegrationAssociationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:DataIntegrationAssociationSummary', dataIntegrationAssociationSummary_clientId - The identifier for teh client that is associated with the DataIntegration association.

$sel:dataIntegrationAssociationArn:DataIntegrationAssociationSummary', dataIntegrationAssociationSummary_dataIntegrationAssociationArn - The Amazon Resource Name (ARN) of the DataIntegration association.

$sel:dataIntegrationArn:DataIntegrationAssociationSummary', dataIntegrationAssociationSummary_dataIntegrationArn - The Amazon Resource Name (ARN)of the DataIntegration.

dataIntegrationAssociationSummary_clientId :: Lens' DataIntegrationAssociationSummary (Maybe Text)

The identifier for teh client that is associated with the DataIntegration association.

dataIntegrationAssociationSummary_dataIntegrationAssociationArn :: Lens' DataIntegrationAssociationSummary (Maybe Text)

The Amazon Resource Name (ARN) of the DataIntegration association.

dataIntegrationAssociationSummary_dataIntegrationArn :: Lens' DataIntegrationAssociationSummary (Maybe Text)

The Amazon Resource Name (ARN)of the DataIntegration.