Network.AWS.AppIntegrationS.Types.EventIntegrationAssociation

Description

Documentation

data EventIntegrationAssociation

The event integration association.

See: newEventIntegrationAssociation smart constructor.

Constructors

EventIntegrationAssociation'

Fields

clientId :: Maybe Text

The identifier for the client that is associated with the event integration.

eventIntegrationName :: Maybe Text

The name of the event integration.

clientAssociationMetadata :: Maybe (HashMap Text Text)

The metadata associated with the client.

eventIntegrationAssociationId :: Maybe Text

The identifier for the event integration association.

eventIntegrationAssociationArn :: Maybe Text

The Amazon Resource Name (ARN) for the event integration association.

eventBridgeRuleName :: Maybe Text

The name of the EventBridge rule.

newEventIntegrationAssociation :: EventIntegrationAssociation

Create a value of EventIntegrationAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:EventIntegrationAssociation', eventIntegrationAssociation_clientId - The identifier for the client that is associated with the event integration.

$sel:eventIntegrationName:EventIntegrationAssociation', eventIntegrationAssociation_eventIntegrationName - The name of the event integration.

$sel:clientAssociationMetadata:EventIntegrationAssociation', eventIntegrationAssociation_clientAssociationMetadata - The metadata associated with the client.

$sel:eventIntegrationAssociationId:EventIntegrationAssociation', eventIntegrationAssociation_eventIntegrationAssociationId - The identifier for the event integration association.

$sel:eventIntegrationAssociationArn:EventIntegrationAssociation', eventIntegrationAssociation_eventIntegrationAssociationArn - The Amazon Resource Name (ARN) for the event integration association.

$sel:eventBridgeRuleName:EventIntegrationAssociation', eventIntegrationAssociation_eventBridgeRuleName - The name of the EventBridge rule.

eventIntegrationAssociation_clientId :: Lens' EventIntegrationAssociation (Maybe Text)

The identifier for the client that is associated with the event integration.

eventIntegrationAssociation_eventIntegrationName :: Lens' EventIntegrationAssociation (Maybe Text)

The name of the event integration.

eventIntegrationAssociation_clientAssociationMetadata :: Lens' EventIntegrationAssociation (Maybe (HashMap Text Text))

The metadata associated with the client.

eventIntegrationAssociation_eventIntegrationAssociationId :: Lens' EventIntegrationAssociation (Maybe Text)

The identifier for the event integration association.

eventIntegrationAssociation_eventIntegrationAssociationArn :: Lens' EventIntegrationAssociation (Maybe Text)

The Amazon Resource Name (ARN) for the event integration association.

eventIntegrationAssociation_eventBridgeRuleName :: Lens' EventIntegrationAssociation (Maybe Text)

The name of the EventBridge rule.