Network.AWS.AppMesh.CreateMesh

Description

Creates a service mesh.

A service mesh is a logical boundary for network traffic between services that are represented by resources within the mesh. After you create your service mesh, you can create virtual services, virtual nodes, virtual routers, and routes to distribute traffic between the applications in your mesh.

For more information about service meshes, see Service meshes.

Creating a Request

data CreateMesh

See: newCreateMesh smart constructor.

Constructors

CreateMesh'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

spec :: Maybe MeshSpec

The service mesh specification to apply.

tags :: Maybe [TagRef]

Optional metadata that you can apply to the service mesh to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

meshName :: Text

The name to use for the service mesh.

newCreateMesh

Arguments

:: Text
$sel:meshName:CreateMesh'

-> CreateMesh

Create a value of CreateMesh with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateMesh', createMesh_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

$sel:spec:CreateMesh', createMesh_spec - The service mesh specification to apply.

$sel:tags:CreateMesh', createMesh_tags - Optional metadata that you can apply to the service mesh to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

$sel:meshName:CreateMesh', createMesh_meshName - The name to use for the service mesh.

Request Lenses

createMesh_clientToken :: Lens' CreateMesh (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

createMesh_spec :: Lens' CreateMesh (Maybe MeshSpec)

The service mesh specification to apply.

createMesh_tags :: Lens' CreateMesh (Maybe [TagRef])

Optional metadata that you can apply to the service mesh to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

createMesh_meshName :: Lens' CreateMesh Text

The name to use for the service mesh.

Destructuring the Response

data CreateMeshResponse

See: newCreateMeshResponse smart constructor.

Constructors

CreateMeshResponse'

Fields

httpStatus :: Int

The response's http status code.

mesh :: MeshData

The full description of your service mesh following the create call.

newCreateMeshResponse

Arguments

:: Int
$sel:httpStatus:CreateMeshResponse'

-> MeshData
$sel:mesh:CreateMeshResponse'

-> CreateMeshResponse

Create a value of CreateMeshResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateMeshResponse', createMeshResponse_httpStatus - The response's http status code.

$sel:mesh:CreateMeshResponse', createMeshResponse_mesh - The full description of your service mesh following the create call.

Response Lenses

createMeshResponse_httpStatus :: Lens' CreateMeshResponse Int

The response's http status code.

createMeshResponse_mesh :: Lens' CreateMeshResponse MeshData

The full description of your service mesh following the create call.