Network.AWS.AppMesh.CreateRoute

Description

Creates a route that is associated with a virtual router.

You can route several different protocols and define a retry policy for a route. Traffic can be routed to one or more virtual nodes.

For more information about routes, see Routes.

Creating a Request

data CreateRoute

See: newCreateRoute smart constructor.

Constructors

CreateRoute'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then the account that you specify must share the mesh with your account before you can create the resource in the service mesh. For more information about mesh sharing, see Working with shared meshes.

tags :: Maybe [TagRef]

Optional metadata that you can apply to the route to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

meshName :: Text

The name of the service mesh to create the route in.

routeName :: Text

The name to use for the route.

spec :: RouteSpec

The route specification to apply.

virtualRouterName :: Text

The name of the virtual router in which to create the route. If the virtual router is in a shared mesh, then you must be the owner of the virtual router resource.

newCreateRoute

Arguments

:: Text
$sel:meshName:CreateRoute'

-> Text
$sel:routeName:CreateRoute'

-> RouteSpec
$sel:spec:CreateRoute'

-> Text
$sel:virtualRouterName:CreateRoute'

-> CreateRoute

Create a value of CreateRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateRoute', createRoute_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

$sel:meshOwner:CreateRoute', createRoute_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then the account that you specify must share the mesh with your account before you can create the resource in the service mesh. For more information about mesh sharing, see Working with shared meshes.

$sel:tags:CreateRoute', createRoute_tags - Optional metadata that you can apply to the route to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

$sel:meshName:CreateRoute', createRoute_meshName - The name of the service mesh to create the route in.

$sel:routeName:CreateRoute', createRoute_routeName - The name to use for the route.

$sel:spec:CreateRoute', createRoute_spec - The route specification to apply.

$sel:virtualRouterName:CreateRoute', createRoute_virtualRouterName - The name of the virtual router in which to create the route. If the virtual router is in a shared mesh, then you must be the owner of the virtual router resource.

Request Lenses

createRoute_clientToken :: Lens' CreateRoute (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

createRoute_meshOwner :: Lens' CreateRoute (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then the account that you specify must share the mesh with your account before you can create the resource in the service mesh. For more information about mesh sharing, see Working with shared meshes.

createRoute_tags :: Lens' CreateRoute (Maybe [TagRef])

Optional metadata that you can apply to the route to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

createRoute_meshName :: Lens' CreateRoute Text

The name of the service mesh to create the route in.

createRoute_routeName :: Lens' CreateRoute Text

The name to use for the route.

createRoute_spec :: Lens' CreateRoute RouteSpec

The route specification to apply.

createRoute_virtualRouterName :: Lens' CreateRoute Text

The name of the virtual router in which to create the route. If the virtual router is in a shared mesh, then you must be the owner of the virtual router resource.

Destructuring the Response

data CreateRouteResponse

See: newCreateRouteResponse smart constructor.

Constructors

CreateRouteResponse'

Fields

httpStatus :: Int

The response's http status code.

route :: RouteData

The full description of your mesh following the create call.

newCreateRouteResponse

Arguments

:: Int
$sel:httpStatus:CreateRouteResponse'

-> RouteData
$sel:route:CreateRouteResponse'

-> CreateRouteResponse

Create a value of CreateRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateRouteResponse', createRouteResponse_httpStatus - The response's http status code.

$sel:route:CreateRouteResponse', createRouteResponse_route - The full description of your mesh following the create call.

Response Lenses

createRouteResponse_httpStatus :: Lens' CreateRouteResponse Int

The response's http status code.

createRouteResponse_route :: Lens' CreateRouteResponse RouteData

The full description of your mesh following the create call.