Network.AWS.AppMesh.CreateVirtualRouter

Description

Creates a virtual router within a service mesh.

Specify a listener for any inbound traffic that your virtual router receives. Create a virtual router for each protocol and port that you need to route. Virtual routers handle traffic for one or more virtual services within your mesh. After you create your virtual router, create and associate routes for your virtual router that direct incoming requests to different virtual nodes.

For more information about virtual routers, see Virtual routers.

Creating a Request

data CreateVirtualRouter

See: newCreateVirtualRouter smart constructor.

Constructors

CreateVirtualRouter'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then the account that you specify must share the mesh with your account before you can create the resource in the service mesh. For more information about mesh sharing, see Working with shared meshes.

tags :: Maybe [TagRef]

Optional metadata that you can apply to the virtual router to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

meshName :: Text

The name of the service mesh to create the virtual router in.

spec :: VirtualRouterSpec

The virtual router specification to apply.

virtualRouterName :: Text

The name to use for the virtual router.

newCreateVirtualRouter

Arguments

:: Text
$sel:meshName:CreateVirtualRouter'

-> VirtualRouterSpec
$sel:spec:CreateVirtualRouter'

-> Text
$sel:virtualRouterName:CreateVirtualRouter'

-> CreateVirtualRouter

Create a value of CreateVirtualRouter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateVirtualRouter', createVirtualRouter_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

$sel:meshOwner:CreateVirtualRouter', createVirtualRouter_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then the account that you specify must share the mesh with your account before you can create the resource in the service mesh. For more information about mesh sharing, see Working with shared meshes.

$sel:tags:CreateVirtualRouter', createVirtualRouter_tags - Optional metadata that you can apply to the virtual router to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

$sel:meshName:CreateVirtualRouter', createVirtualRouter_meshName - The name of the service mesh to create the virtual router in.

$sel:spec:CreateVirtualRouter', createVirtualRouter_spec - The virtual router specification to apply.

$sel:virtualRouterName:CreateVirtualRouter', createVirtualRouter_virtualRouterName - The name to use for the virtual router.

Request Lenses

createVirtualRouter_clientToken :: Lens' CreateVirtualRouter (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

createVirtualRouter_meshOwner :: Lens' CreateVirtualRouter (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then the account that you specify must share the mesh with your account before you can create the resource in the service mesh. For more information about mesh sharing, see Working with shared meshes.

createVirtualRouter_tags :: Lens' CreateVirtualRouter (Maybe [TagRef])

Optional metadata that you can apply to the virtual router to assist with categorization and organization. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

createVirtualRouter_meshName :: Lens' CreateVirtualRouter Text

The name of the service mesh to create the virtual router in.

createVirtualRouter_spec :: Lens' CreateVirtualRouter VirtualRouterSpec

The virtual router specification to apply.

createVirtualRouter_virtualRouterName :: Lens' CreateVirtualRouter Text

The name to use for the virtual router.

Destructuring the Response

data CreateVirtualRouterResponse

See: newCreateVirtualRouterResponse smart constructor.

Constructors

CreateVirtualRouterResponse'

Fields

httpStatus :: Int

The response's http status code.

virtualRouter :: VirtualRouterData

The full description of your virtual router following the create call.

newCreateVirtualRouterResponse

Arguments

:: Int
$sel:httpStatus:CreateVirtualRouterResponse'

-> VirtualRouterData
$sel:virtualRouter:CreateVirtualRouterResponse'

-> CreateVirtualRouterResponse

Create a value of CreateVirtualRouterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateVirtualRouterResponse', createVirtualRouterResponse_httpStatus - The response's http status code.

$sel:virtualRouter:CreateVirtualRouterResponse', createVirtualRouterResponse_virtualRouter - The full description of your virtual router following the create call.

Response Lenses

createVirtualRouterResponse_httpStatus :: Lens' CreateVirtualRouterResponse Int

The response's http status code.

createVirtualRouterResponse_virtualRouter :: Lens' CreateVirtualRouterResponse VirtualRouterData

The full description of your virtual router following the create call.