Network.AWS.AppMesh.DeleteGatewayRoute

Description

Deletes an existing gateway route.

Creating a Request

data DeleteGatewayRoute

See: newDeleteGatewayRoute smart constructor.

Constructors

DeleteGatewayRoute'

Fields

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

gatewayRouteName :: Text

The name of the gateway route to delete.

meshName :: Text

The name of the service mesh to delete the gateway route from.

virtualGatewayName :: Text

The name of the virtual gateway to delete the route from.

newDeleteGatewayRoute

Arguments

:: Text
$sel:gatewayRouteName:DeleteGatewayRoute'

-> Text
$sel:meshName:DeleteGatewayRoute'

-> Text
$sel:virtualGatewayName:DeleteGatewayRoute'

-> DeleteGatewayRoute

Create a value of DeleteGatewayRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshOwner:DeleteGatewayRoute', deleteGatewayRoute_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:gatewayRouteName:DeleteGatewayRoute', deleteGatewayRoute_gatewayRouteName - The name of the gateway route to delete.

$sel:meshName:DeleteGatewayRoute', deleteGatewayRoute_meshName - The name of the service mesh to delete the gateway route from.

$sel:virtualGatewayName:DeleteGatewayRoute', deleteGatewayRoute_virtualGatewayName - The name of the virtual gateway to delete the route from.

Request Lenses

deleteGatewayRoute_meshOwner :: Lens' DeleteGatewayRoute (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

deleteGatewayRoute_gatewayRouteName :: Lens' DeleteGatewayRoute Text

The name of the gateway route to delete.

deleteGatewayRoute_meshName :: Lens' DeleteGatewayRoute Text

The name of the service mesh to delete the gateway route from.

deleteGatewayRoute_virtualGatewayName :: Lens' DeleteGatewayRoute Text

The name of the virtual gateway to delete the route from.

Destructuring the Response

data DeleteGatewayRouteResponse

See: newDeleteGatewayRouteResponse smart constructor.

Constructors

DeleteGatewayRouteResponse'

Fields

httpStatus :: Int

The response's http status code.

gatewayRoute :: GatewayRouteData

The gateway route that was deleted.

newDeleteGatewayRouteResponse

Arguments

:: Int
$sel:httpStatus:DeleteGatewayRouteResponse'

-> GatewayRouteData
$sel:gatewayRoute:DeleteGatewayRouteResponse'

-> DeleteGatewayRouteResponse

Create a value of DeleteGatewayRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteGatewayRouteResponse', deleteGatewayRouteResponse_httpStatus - The response's http status code.

$sel:gatewayRoute:DeleteGatewayRouteResponse', deleteGatewayRouteResponse_gatewayRoute - The gateway route that was deleted.

Response Lenses

deleteGatewayRouteResponse_httpStatus :: Lens' DeleteGatewayRouteResponse Int

The response's http status code.

deleteGatewayRouteResponse_gatewayRoute :: Lens' DeleteGatewayRouteResponse GatewayRouteData

The gateway route that was deleted.