Network.AWS.AppMesh.DeleteVirtualNode

Description

Deletes an existing virtual node.

You must delete any virtual services that list a virtual node as a service provider before you can delete the virtual node itself.

Creating a Request

data DeleteVirtualNode

Deletes a virtual node input.

See: newDeleteVirtualNode smart constructor.

Constructors

DeleteVirtualNode'

Fields

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

meshName :: Text

The name of the service mesh to delete the virtual node in.

virtualNodeName :: Text

The name of the virtual node to delete.

newDeleteVirtualNode

Arguments

:: Text
$sel:meshName:DeleteVirtualNode'

-> Text
$sel:virtualNodeName:DeleteVirtualNode'

-> DeleteVirtualNode

Create a value of DeleteVirtualNode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshOwner:DeleteVirtualNode', deleteVirtualNode_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:meshName:DeleteVirtualNode', deleteVirtualNode_meshName - The name of the service mesh to delete the virtual node in.

$sel:virtualNodeName:DeleteVirtualNode', deleteVirtualNode_virtualNodeName - The name of the virtual node to delete.

Request Lenses

deleteVirtualNode_meshOwner :: Lens' DeleteVirtualNode (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

deleteVirtualNode_meshName :: Lens' DeleteVirtualNode Text

The name of the service mesh to delete the virtual node in.

deleteVirtualNode_virtualNodeName :: Lens' DeleteVirtualNode Text

The name of the virtual node to delete.

Destructuring the Response

data DeleteVirtualNodeResponse

See: newDeleteVirtualNodeResponse smart constructor.

Constructors

DeleteVirtualNodeResponse'

Fields

httpStatus :: Int

The response's http status code.

virtualNode :: VirtualNodeData

The virtual node that was deleted.

newDeleteVirtualNodeResponse

Arguments

:: Int
$sel:httpStatus:DeleteVirtualNodeResponse'

-> VirtualNodeData
$sel:virtualNode:DeleteVirtualNodeResponse'

-> DeleteVirtualNodeResponse

Create a value of DeleteVirtualNodeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteVirtualNodeResponse', deleteVirtualNodeResponse_httpStatus - The response's http status code.

$sel:virtualNode:DeleteVirtualNodeResponse', deleteVirtualNodeResponse_virtualNode - The virtual node that was deleted.

Response Lenses

deleteVirtualNodeResponse_httpStatus :: Lens' DeleteVirtualNodeResponse Int

The response's http status code.

deleteVirtualNodeResponse_virtualNode :: Lens' DeleteVirtualNodeResponse VirtualNodeData

The virtual node that was deleted.