Network.AWS.AppMesh.DescribeGatewayRoute

Description

Describes an existing gateway route.

Creating a Request

data DescribeGatewayRoute

See: newDescribeGatewayRoute smart constructor.

Constructors

DescribeGatewayRoute'

Fields

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

gatewayRouteName :: Text

The name of the gateway route to describe.

meshName :: Text

The name of the service mesh that the gateway route resides in.

virtualGatewayName :: Text

The name of the virtual gateway that the gateway route is associated with.

newDescribeGatewayRoute

Arguments

:: Text
$sel:gatewayRouteName:DescribeGatewayRoute'

-> Text
$sel:meshName:DescribeGatewayRoute'

-> Text
$sel:virtualGatewayName:DescribeGatewayRoute'

-> DescribeGatewayRoute

Create a value of DescribeGatewayRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshOwner:DescribeGatewayRoute', describeGatewayRoute_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:gatewayRouteName:DescribeGatewayRoute', describeGatewayRoute_gatewayRouteName - The name of the gateway route to describe.

$sel:meshName:DescribeGatewayRoute', describeGatewayRoute_meshName - The name of the service mesh that the gateway route resides in.

$sel:virtualGatewayName:DescribeGatewayRoute', describeGatewayRoute_virtualGatewayName - The name of the virtual gateway that the gateway route is associated with.

Request Lenses

describeGatewayRoute_meshOwner :: Lens' DescribeGatewayRoute (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

describeGatewayRoute_gatewayRouteName :: Lens' DescribeGatewayRoute Text

The name of the gateway route to describe.

describeGatewayRoute_meshName :: Lens' DescribeGatewayRoute Text

The name of the service mesh that the gateway route resides in.

describeGatewayRoute_virtualGatewayName :: Lens' DescribeGatewayRoute Text

The name of the virtual gateway that the gateway route is associated with.

Destructuring the Response

data DescribeGatewayRouteResponse

See: newDescribeGatewayRouteResponse smart constructor.

Constructors

DescribeGatewayRouteResponse'

Fields

httpStatus :: Int

The response's http status code.

gatewayRoute :: GatewayRouteData

The full description of your gateway route.

newDescribeGatewayRouteResponse

Arguments

:: Int
$sel:httpStatus:DescribeGatewayRouteResponse'

-> GatewayRouteData
$sel:gatewayRoute:DescribeGatewayRouteResponse'

-> DescribeGatewayRouteResponse

Create a value of DescribeGatewayRouteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeGatewayRouteResponse', describeGatewayRouteResponse_httpStatus - The response's http status code.

$sel:gatewayRoute:DescribeGatewayRouteResponse', describeGatewayRouteResponse_gatewayRoute - The full description of your gateway route.

Response Lenses

describeGatewayRouteResponse_httpStatus :: Lens' DescribeGatewayRouteResponse Int

The response's http status code.

describeGatewayRouteResponse_gatewayRoute :: Lens' DescribeGatewayRouteResponse GatewayRouteData

The full description of your gateway route.