Network.AWS.AppMesh.ListVirtualServices

Description

Returns a list of existing virtual services in a service mesh.

This operation returns paginated results.

Creating a Request

data ListVirtualServices

See: newListVirtualServices smart constructor.

Constructors

ListVirtualServices'

Fields

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

nextToken :: Maybe Text

The nextToken value returned from a previous paginated ListVirtualServices request where limit was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

limit :: Maybe Natural

The maximum number of results returned by ListVirtualServices in paginated output. When you use this parameter, ListVirtualServices returns only limit results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListVirtualServices request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListVirtualServices returns up to 100 results and a nextToken value if applicable.

meshName :: Text

The name of the service mesh to list virtual services in.

newListVirtualServices

Arguments

:: Text
$sel:meshName:ListVirtualServices'

-> ListVirtualServices

Create a value of ListVirtualServices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshOwner:ListVirtualServices', listVirtualServices_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:nextToken:ListVirtualServices', listVirtualServices_nextToken - The nextToken value returned from a previous paginated ListVirtualServices request where limit was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

$sel:limit:ListVirtualServices', listVirtualServices_limit - The maximum number of results returned by ListVirtualServices in paginated output. When you use this parameter, ListVirtualServices returns only limit results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListVirtualServices request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListVirtualServices returns up to 100 results and a nextToken value if applicable.

$sel:meshName:ListVirtualServices', listVirtualServices_meshName - The name of the service mesh to list virtual services in.

Request Lenses

listVirtualServices_meshOwner :: Lens' ListVirtualServices (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

listVirtualServices_nextToken :: Lens' ListVirtualServices (Maybe Text)

The nextToken value returned from a previous paginated ListVirtualServices request where limit was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value.

listVirtualServices_limit :: Lens' ListVirtualServices (Maybe Natural)

The maximum number of results returned by ListVirtualServices in paginated output. When you use this parameter, ListVirtualServices returns only limit results in a single page along with a nextToken response element. You can see the remaining results of the initial request by sending another ListVirtualServices request with the returned nextToken value. This value can be between 1 and 100. If you don't use this parameter, ListVirtualServices returns up to 100 results and a nextToken value if applicable.

listVirtualServices_meshName :: Lens' ListVirtualServices Text

The name of the service mesh to list virtual services in.

Destructuring the Response

data ListVirtualServicesResponse

See: newListVirtualServicesResponse smart constructor.

Constructors

ListVirtualServicesResponse'

Fields

nextToken :: Maybe Text

The nextToken value to include in a future ListVirtualServices request. When the results of a ListVirtualServices request exceed limit, you can use this value to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

virtualServices :: [VirtualServiceRef]

The list of existing virtual services for the specified service mesh.

newListVirtualServicesResponse

Arguments

:: Int
$sel:httpStatus:ListVirtualServicesResponse'

-> ListVirtualServicesResponse

Create a value of ListVirtualServicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListVirtualServices', listVirtualServicesResponse_nextToken - The nextToken value to include in a future ListVirtualServices request. When the results of a ListVirtualServices request exceed limit, you can use this value to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListVirtualServicesResponse', listVirtualServicesResponse_httpStatus - The response's http status code.

$sel:virtualServices:ListVirtualServicesResponse', listVirtualServicesResponse_virtualServices - The list of existing virtual services for the specified service mesh.

Response Lenses

listVirtualServicesResponse_nextToken :: Lens' ListVirtualServicesResponse (Maybe Text)

The nextToken value to include in a future ListVirtualServices request. When the results of a ListVirtualServices request exceed limit, you can use this value to retrieve the next page of results. This value is null when there are no more results to return.

listVirtualServicesResponse_httpStatus :: Lens' ListVirtualServicesResponse Int

The response's http status code.

listVirtualServicesResponse_virtualServices :: Lens' ListVirtualServicesResponse [VirtualServiceRef]

The list of existing virtual services for the specified service mesh.