Network.AWS.AppMesh.Types.GatewayRouteRef

Description

Documentation

data GatewayRouteRef

An object that represents a gateway route returned by a list operation.

See: newGatewayRouteRef smart constructor.

Constructors

GatewayRouteRef'

Fields

arn :: Text

The full Amazon Resource Name (ARN) for the gateway route.

createdAt :: POSIX

The Unix epoch timestamp in seconds for when the resource was created.

gatewayRouteName :: Text

The name of the gateway route.

lastUpdatedAt :: POSIX

The Unix epoch timestamp in seconds for when the resource was last updated.

meshName :: Text

The name of the service mesh that the resource resides in.

meshOwner :: Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

resourceOwner :: Text

The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

version :: Integer

The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

virtualGatewayName :: Text

The virtual gateway that the gateway route is associated with.

newGatewayRouteRef

Arguments

:: Text
$sel:arn:GatewayRouteRef'

-> UTCTime
$sel:createdAt:GatewayRouteRef'

-> Text
$sel:gatewayRouteName:GatewayRouteRef'

-> UTCTime
$sel:lastUpdatedAt:GatewayRouteRef'

-> Text
$sel:meshName:GatewayRouteRef'

-> Text
$sel:meshOwner:GatewayRouteRef'

-> Text
$sel:resourceOwner:GatewayRouteRef'

-> Integer
$sel:version:GatewayRouteRef'

-> Text
$sel:virtualGatewayName:GatewayRouteRef'

-> GatewayRouteRef

Create a value of GatewayRouteRef with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GatewayRouteRef', gatewayRouteRef_arn - The full Amazon Resource Name (ARN) for the gateway route.

$sel:createdAt:GatewayRouteRef', gatewayRouteRef_createdAt - The Unix epoch timestamp in seconds for when the resource was created.

$sel:gatewayRouteName:GatewayRouteRef', gatewayRouteRef_gatewayRouteName - The name of the gateway route.

$sel:lastUpdatedAt:GatewayRouteRef', gatewayRouteRef_lastUpdatedAt - The Unix epoch timestamp in seconds for when the resource was last updated.

$sel:meshName:GatewayRouteRef', gatewayRouteRef_meshName - The name of the service mesh that the resource resides in.

$sel:meshOwner:GatewayRouteRef', gatewayRouteRef_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:resourceOwner:GatewayRouteRef', gatewayRouteRef_resourceOwner - The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

$sel:version:GatewayRouteRef', gatewayRouteRef_version - The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

$sel:virtualGatewayName:GatewayRouteRef', gatewayRouteRef_virtualGatewayName - The virtual gateway that the gateway route is associated with.

gatewayRouteRef_arn :: Lens' GatewayRouteRef Text

The full Amazon Resource Name (ARN) for the gateway route.

gatewayRouteRef_createdAt :: Lens' GatewayRouteRef UTCTime

The Unix epoch timestamp in seconds for when the resource was created.

gatewayRouteRef_gatewayRouteName :: Lens' GatewayRouteRef Text

The name of the gateway route.

gatewayRouteRef_lastUpdatedAt :: Lens' GatewayRouteRef UTCTime

The Unix epoch timestamp in seconds for when the resource was last updated.

gatewayRouteRef_meshName :: Lens' GatewayRouteRef Text

The name of the service mesh that the resource resides in.

gatewayRouteRef_meshOwner :: Lens' GatewayRouteRef Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

gatewayRouteRef_resourceOwner :: Lens' GatewayRouteRef Text

The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

gatewayRouteRef_version :: Lens' GatewayRouteRef Integer

The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

gatewayRouteRef_virtualGatewayName :: Lens' GatewayRouteRef Text

The virtual gateway that the gateway route is associated with.