Network.AWS.AppMesh.Types.GrpcRoute

Description

Documentation

data GrpcRoute

An object that represents a gRPC route type.

See: newGrpcRoute smart constructor.

Constructors

GrpcRoute'

Fields

retryPolicy :: Maybe GrpcRetryPolicy

An object that represents a retry policy.

timeout :: Maybe GrpcTimeout

An object that represents types of timeouts.

action :: GrpcRouteAction

An object that represents the action to take if a match is determined.

match :: GrpcRouteMatch

An object that represents the criteria for determining a request match.

newGrpcRoute

Arguments

:: GrpcRouteAction
$sel:action:GrpcRoute'

-> GrpcRouteMatch
$sel:match:GrpcRoute'

-> GrpcRoute

Create a value of GrpcRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retryPolicy:GrpcRoute', grpcRoute_retryPolicy - An object that represents a retry policy.

$sel:timeout:GrpcRoute', grpcRoute_timeout - An object that represents types of timeouts.

$sel:action:GrpcRoute', grpcRoute_action - An object that represents the action to take if a match is determined.

$sel:match:GrpcRoute', grpcRoute_match - An object that represents the criteria for determining a request match.

grpcRoute_retryPolicy :: Lens' GrpcRoute (Maybe GrpcRetryPolicy)

An object that represents a retry policy.

grpcRoute_timeout :: Lens' GrpcRoute (Maybe GrpcTimeout)

An object that represents types of timeouts.

grpcRoute_action :: Lens' GrpcRoute GrpcRouteAction

An object that represents the action to take if a match is determined.

grpcRoute_match :: Lens' GrpcRoute GrpcRouteMatch

An object that represents the criteria for determining a request match.