Network.AWS.AppMesh.Types.HttpGatewayRoute

Description

Documentation

data HttpGatewayRoute

An object that represents an HTTP gateway route.

See: newHttpGatewayRoute smart constructor.

Constructors

HttpGatewayRoute'

Fields

action :: HttpGatewayRouteAction

An object that represents the action to take if a match is determined.

match :: HttpGatewayRouteMatch

An object that represents the criteria for determining a request match.

newHttpGatewayRoute

Arguments

:: HttpGatewayRouteAction
$sel:action:HttpGatewayRoute'

-> HttpGatewayRouteMatch
$sel:match:HttpGatewayRoute'

-> HttpGatewayRoute

Create a value of HttpGatewayRoute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:HttpGatewayRoute', httpGatewayRoute_action - An object that represents the action to take if a match is determined.

$sel:match:HttpGatewayRoute', httpGatewayRoute_match - An object that represents the criteria for determining a request match.

httpGatewayRoute_action :: Lens' HttpGatewayRoute HttpGatewayRouteAction

An object that represents the action to take if a match is determined.

httpGatewayRoute_match :: Lens' HttpGatewayRoute HttpGatewayRouteMatch

An object that represents the criteria for determining a request match.