Network.AWS.AppMesh.Types.RouteData

Description

Documentation

data RouteData

An object that represents a route returned by a describe operation.

See: newRouteData smart constructor.

Constructors

RouteData'

Fields

meshName :: Text

The name of the service mesh that the route resides in.

metadata :: ResourceMetadata

The associated metadata for the route.

routeName :: Text

The name of the route.

spec :: RouteSpec

The specifications of the route.

status :: RouteStatus

The status of the route.

virtualRouterName :: Text

The virtual router that the route is associated with.

newRouteData

Arguments

:: Text
$sel:meshName:RouteData'

-> ResourceMetadata
$sel:metadata:RouteData'

-> Text
$sel:routeName:RouteData'

-> RouteSpec
$sel:spec:RouteData'

-> RouteStatus
$sel:status:RouteData'

-> Text
$sel:virtualRouterName:RouteData'

-> RouteData

Create a value of RouteData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshName:RouteData', routeData_meshName - The name of the service mesh that the route resides in.

$sel:metadata:RouteData', routeData_metadata - The associated metadata for the route.

$sel:routeName:RouteData', routeData_routeName - The name of the route.

$sel:spec:RouteData', routeData_spec - The specifications of the route.

$sel:status:RouteData', routeData_status - The status of the route.

$sel:virtualRouterName:RouteData', routeData_virtualRouterName - The virtual router that the route is associated with.

routeData_meshName :: Lens' RouteData Text

The name of the service mesh that the route resides in.

routeData_metadata :: Lens' RouteData ResourceMetadata

The associated metadata for the route.

routeData_routeName :: Lens' RouteData Text

The name of the route.

routeData_spec :: Lens' RouteData RouteSpec

The specifications of the route.

routeData_status :: Lens' RouteData RouteStatus

The status of the route.

routeData_virtualRouterName :: Lens' RouteData Text

The virtual router that the route is associated with.