Network.AWS.AppMesh.Types.RouteRef

Description

Documentation

data RouteRef

An object that represents a route returned by a list operation.

See: newRouteRef smart constructor.

Constructors

RouteRef'

Fields

arn :: Text

The full Amazon Resource Name (ARN) for the route.

createdAt :: POSIX

The Unix epoch timestamp in seconds for when the resource was created.

lastUpdatedAt :: POSIX

The Unix epoch timestamp in seconds for when the resource was last updated.

meshName :: Text

The name of the service mesh that the route resides in.

meshOwner :: Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

resourceOwner :: Text

The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

routeName :: Text

The name of the route.

version :: Integer

The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

virtualRouterName :: Text

The virtual router that the route is associated with.

newRouteRef

Arguments

:: Text
$sel:arn:RouteRef'

-> UTCTime
$sel:createdAt:RouteRef'

-> UTCTime
$sel:lastUpdatedAt:RouteRef'

-> Text
$sel:meshName:RouteRef'

-> Text
$sel:meshOwner:RouteRef'

-> Text
$sel:resourceOwner:RouteRef'

-> Text
$sel:routeName:RouteRef'

-> Integer
$sel:version:RouteRef'

-> Text
$sel:virtualRouterName:RouteRef'

-> RouteRef

Create a value of RouteRef with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:RouteRef', routeRef_arn - The full Amazon Resource Name (ARN) for the route.

$sel:createdAt:RouteRef', routeRef_createdAt - The Unix epoch timestamp in seconds for when the resource was created.

$sel:lastUpdatedAt:RouteRef', routeRef_lastUpdatedAt - The Unix epoch timestamp in seconds for when the resource was last updated.

$sel:meshName:RouteRef', routeRef_meshName - The name of the service mesh that the route resides in.

$sel:meshOwner:RouteRef', routeRef_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:resourceOwner:RouteRef', routeRef_resourceOwner - The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

$sel:routeName:RouteRef', routeRef_routeName - The name of the route.

$sel:version:RouteRef', routeRef_version - The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

$sel:virtualRouterName:RouteRef', routeRef_virtualRouterName - The virtual router that the route is associated with.

routeRef_arn :: Lens' RouteRef Text

The full Amazon Resource Name (ARN) for the route.

routeRef_createdAt :: Lens' RouteRef UTCTime

The Unix epoch timestamp in seconds for when the resource was created.

routeRef_lastUpdatedAt :: Lens' RouteRef UTCTime

The Unix epoch timestamp in seconds for when the resource was last updated.

routeRef_meshName :: Lens' RouteRef Text

The name of the service mesh that the route resides in.

routeRef_meshOwner :: Lens' RouteRef Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

routeRef_resourceOwner :: Lens' RouteRef Text

The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

routeRef_routeName :: Lens' RouteRef Text

The name of the route.

routeRef_version :: Lens' RouteRef Integer

The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

routeRef_virtualRouterName :: Lens' RouteRef Text

The virtual router that the route is associated with.