Network.AWS.AppMesh.Types.RouteSpec

Description

Documentation

data RouteSpec

An object that represents a route specification. Specify one route type.

See: newRouteSpec smart constructor.

Constructors

RouteSpec'

Fields

priority :: Maybe Natural

The priority for the route. Routes are matched based on the specified value, where 0 is the highest priority.

http2Route :: Maybe HttpRoute

An object that represents the specification of an HTTP/2 route.

grpcRoute :: Maybe GrpcRoute

An object that represents the specification of a gRPC route.

tcpRoute :: Maybe TcpRoute

An object that represents the specification of a TCP route.

httpRoute :: Maybe HttpRoute

An object that represents the specification of an HTTP route.

newRouteSpec :: RouteSpec

Create a value of RouteSpec with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:priority:RouteSpec', routeSpec_priority - The priority for the route. Routes are matched based on the specified value, where 0 is the highest priority.

$sel:http2Route:RouteSpec', routeSpec_http2Route - An object that represents the specification of an HTTP/2 route.

$sel:grpcRoute:RouteSpec', routeSpec_grpcRoute - An object that represents the specification of a gRPC route.

$sel:tcpRoute:RouteSpec', routeSpec_tcpRoute - An object that represents the specification of a TCP route.

$sel:httpRoute:RouteSpec', routeSpec_httpRoute - An object that represents the specification of an HTTP route.

routeSpec_priority :: Lens' RouteSpec (Maybe Natural)

The priority for the route. Routes are matched based on the specified value, where 0 is the highest priority.

routeSpec_http2Route :: Lens' RouteSpec (Maybe HttpRoute)

An object that represents the specification of an HTTP/2 route.

routeSpec_grpcRoute :: Lens' RouteSpec (Maybe GrpcRoute)

An object that represents the specification of a gRPC route.

routeSpec_tcpRoute :: Lens' RouteSpec (Maybe TcpRoute)

An object that represents the specification of a TCP route.

routeSpec_httpRoute :: Lens' RouteSpec (Maybe HttpRoute)

An object that represents the specification of an HTTP route.