Network.AWS.AppMesh.Types.RouteStatus

Description

Documentation

data RouteStatus

An object that represents the current status of a route.

See: newRouteStatus smart constructor.

Constructors

RouteStatus'

Fields

status :: RouteStatusCode

The current status for the route.

newRouteStatus

Arguments

:: RouteStatusCode
$sel:status:RouteStatus'

-> RouteStatus

Create a value of RouteStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RouteStatus', routeStatus_status - The current status for the route.

routeStatus_status :: Lens' RouteStatus RouteStatusCode

The current status for the route.