Network.AWS.AppMesh.Types.TlsValidationContextTrust

Description

Documentation

data TlsValidationContextTrust

An object that represents a Transport Layer Security (TLS) validation context trust.

See: newTlsValidationContextTrust smart constructor.

Constructors

TlsValidationContextTrust'

Fields

acm :: Maybe TlsValidationContextAcmTrust

A reference to an object that represents a Transport Layer Security (TLS) validation context trust for an Certificate Manager certificate.

sds :: Maybe TlsValidationContextSdsTrust

A reference to an object that represents a Transport Layer Security (TLS) Secret Discovery Service validation context trust.

file :: Maybe TlsValidationContextFileTrust

An object that represents a Transport Layer Security (TLS) validation context trust for a local file.

newTlsValidationContextTrust :: TlsValidationContextTrust

Create a value of TlsValidationContextTrust with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acm:TlsValidationContextTrust', tlsValidationContextTrust_acm - A reference to an object that represents a Transport Layer Security (TLS) validation context trust for an Certificate Manager certificate.

$sel:sds:TlsValidationContextTrust', tlsValidationContextTrust_sds - A reference to an object that represents a Transport Layer Security (TLS) Secret Discovery Service validation context trust.

$sel:file:TlsValidationContextTrust', tlsValidationContextTrust_file - An object that represents a Transport Layer Security (TLS) validation context trust for a local file.

tlsValidationContextTrust_acm :: Lens' TlsValidationContextTrust (Maybe TlsValidationContextAcmTrust)

A reference to an object that represents a Transport Layer Security (TLS) validation context trust for an Certificate Manager certificate.

tlsValidationContextTrust_sds :: Lens' TlsValidationContextTrust (Maybe TlsValidationContextSdsTrust)

A reference to an object that represents a Transport Layer Security (TLS) Secret Discovery Service validation context trust.

tlsValidationContextTrust_file :: Lens' TlsValidationContextTrust (Maybe TlsValidationContextFileTrust)

An object that represents a Transport Layer Security (TLS) validation context trust for a local file.