Network.AWS.AppMesh.Types.VirtualGatewayData

Description

Documentation

data VirtualGatewayData

An object that represents a virtual gateway returned by a describe operation.

See: newVirtualGatewayData smart constructor.

Constructors

VirtualGatewayData'

Fields

meshName :: Text

The name of the service mesh that the virtual gateway resides in.

metadata :: ResourceMetadata

spec :: VirtualGatewaySpec

The specifications of the virtual gateway.

status :: VirtualGatewayStatus

The current status of the virtual gateway.

virtualGatewayName :: Text

The name of the virtual gateway.

newVirtualGatewayData

Arguments

:: Text
$sel:meshName:VirtualGatewayData'

-> ResourceMetadata
$sel:metadata:VirtualGatewayData'

-> VirtualGatewaySpec
$sel:spec:VirtualGatewayData'

-> VirtualGatewayStatus
$sel:status:VirtualGatewayData'

-> Text
$sel:virtualGatewayName:VirtualGatewayData'

-> VirtualGatewayData

Create a value of VirtualGatewayData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meshName:VirtualGatewayData', virtualGatewayData_meshName - The name of the service mesh that the virtual gateway resides in.

$sel:metadata:VirtualGatewayData', virtualGatewayData_metadata - Undocumented member.

$sel:spec:VirtualGatewayData', virtualGatewayData_spec - The specifications of the virtual gateway.

$sel:status:VirtualGatewayData', virtualGatewayData_status - The current status of the virtual gateway.

$sel:virtualGatewayName:VirtualGatewayData', virtualGatewayData_virtualGatewayName - The name of the virtual gateway.

virtualGatewayData_meshName :: Lens' VirtualGatewayData Text

The name of the service mesh that the virtual gateway resides in.

virtualGatewayData_metadata :: Lens' VirtualGatewayData ResourceMetadata

Undocumented member.

virtualGatewayData_spec :: Lens' VirtualGatewayData VirtualGatewaySpec

The specifications of the virtual gateway.

virtualGatewayData_status :: Lens' VirtualGatewayData VirtualGatewayStatus

The current status of the virtual gateway.

virtualGatewayData_virtualGatewayName :: Lens' VirtualGatewayData Text

The name of the virtual gateway.