Network.AWS.AppMesh.Types.VirtualGatewayListener

Description

Documentation

data VirtualGatewayListener

An object that represents a listener for a virtual gateway.

See: newVirtualGatewayListener smart constructor.

Constructors

VirtualGatewayListener'

Fields

healthCheck :: Maybe VirtualGatewayHealthCheckPolicy

The health check information for the listener.

connectionPool :: Maybe VirtualGatewayConnectionPool

The connection pool information for the virtual gateway listener.

tls :: Maybe VirtualGatewayListenerTls

A reference to an object that represents the Transport Layer Security (TLS) properties for the listener.

portMapping :: VirtualGatewayPortMapping

The port mapping information for the listener.

newVirtualGatewayListener

Arguments

:: VirtualGatewayPortMapping
$sel:portMapping:VirtualGatewayListener'

-> VirtualGatewayListener

Create a value of VirtualGatewayListener with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:healthCheck:VirtualGatewayListener', virtualGatewayListener_healthCheck - The health check information for the listener.

$sel:connectionPool:VirtualGatewayListener', virtualGatewayListener_connectionPool - The connection pool information for the virtual gateway listener.

$sel:tls:VirtualGatewayListener', virtualGatewayListener_tls - A reference to an object that represents the Transport Layer Security (TLS) properties for the listener.

$sel:portMapping:VirtualGatewayListener', virtualGatewayListener_portMapping - The port mapping information for the listener.

virtualGatewayListener_healthCheck :: Lens' VirtualGatewayListener (Maybe VirtualGatewayHealthCheckPolicy)

The health check information for the listener.

virtualGatewayListener_connectionPool :: Lens' VirtualGatewayListener (Maybe VirtualGatewayConnectionPool)

The connection pool information for the virtual gateway listener.

virtualGatewayListener_tls :: Lens' VirtualGatewayListener (Maybe VirtualGatewayListenerTls)

A reference to an object that represents the Transport Layer Security (TLS) properties for the listener.

virtualGatewayListener_portMapping :: Lens' VirtualGatewayListener VirtualGatewayPortMapping

The port mapping information for the listener.