Network.AWS.AppMesh.Types.VirtualNodeRef

Description

Documentation

data VirtualNodeRef

An object that represents a virtual node returned by a list operation.

See: newVirtualNodeRef smart constructor.

Constructors

VirtualNodeRef'

Fields

arn :: Text

The full Amazon Resource Name (ARN) for the virtual node.

createdAt :: POSIX

The Unix epoch timestamp in seconds for when the resource was created.

lastUpdatedAt :: POSIX

The Unix epoch timestamp in seconds for when the resource was last updated.

meshName :: Text

The name of the service mesh that the virtual node resides in.

meshOwner :: Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

resourceOwner :: Text

The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

version :: Integer

The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

virtualNodeName :: Text

The name of the virtual node.

newVirtualNodeRef

Arguments

:: Text
$sel:arn:VirtualNodeRef'

-> UTCTime
$sel:createdAt:VirtualNodeRef'

-> UTCTime
$sel:lastUpdatedAt:VirtualNodeRef'

-> Text
$sel:meshName:VirtualNodeRef'

-> Text
$sel:meshOwner:VirtualNodeRef'

-> Text
$sel:resourceOwner:VirtualNodeRef'

-> Integer
$sel:version:VirtualNodeRef'

-> Text
$sel:virtualNodeName:VirtualNodeRef'

-> VirtualNodeRef

Create a value of VirtualNodeRef with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:VirtualNodeRef', virtualNodeRef_arn - The full Amazon Resource Name (ARN) for the virtual node.

$sel:createdAt:VirtualNodeRef', virtualNodeRef_createdAt - The Unix epoch timestamp in seconds for when the resource was created.

$sel:lastUpdatedAt:VirtualNodeRef', virtualNodeRef_lastUpdatedAt - The Unix epoch timestamp in seconds for when the resource was last updated.

$sel:meshName:VirtualNodeRef', virtualNodeRef_meshName - The name of the service mesh that the virtual node resides in.

$sel:meshOwner:VirtualNodeRef', virtualNodeRef_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:resourceOwner:VirtualNodeRef', virtualNodeRef_resourceOwner - The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

$sel:version:VirtualNodeRef', virtualNodeRef_version - The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

$sel:virtualNodeName:VirtualNodeRef', virtualNodeRef_virtualNodeName - The name of the virtual node.

virtualNodeRef_arn :: Lens' VirtualNodeRef Text

The full Amazon Resource Name (ARN) for the virtual node.

virtualNodeRef_createdAt :: Lens' VirtualNodeRef UTCTime

The Unix epoch timestamp in seconds for when the resource was created.

virtualNodeRef_lastUpdatedAt :: Lens' VirtualNodeRef UTCTime

The Unix epoch timestamp in seconds for when the resource was last updated.

virtualNodeRef_meshName :: Lens' VirtualNodeRef Text

The name of the service mesh that the virtual node resides in.

virtualNodeRef_meshOwner :: Lens' VirtualNodeRef Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

virtualNodeRef_resourceOwner :: Lens' VirtualNodeRef Text

The AWS IAM account ID of the resource owner. If the account ID is not your own, then it's the ID of the mesh owner or of another account that the mesh is shared with. For more information about mesh sharing, see Working with shared meshes.

virtualNodeRef_version :: Lens' VirtualNodeRef Integer

The version of the resource. Resources are created at version 1, and this version is incremented each time that they're updated.

virtualNodeRef_virtualNodeName :: Lens' VirtualNodeRef Text

The name of the virtual node.