Network.AWS.AppMesh.UpdateVirtualNode

Description

Updates an existing virtual node in a specified service mesh.

Creating a Request

data UpdateVirtualNode

See: newUpdateVirtualNode smart constructor.

Constructors

UpdateVirtualNode'

Fields

clientToken :: Maybe Text

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

meshOwner :: Maybe Text

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

meshName :: Text

The name of the service mesh that the virtual node resides in.

spec :: VirtualNodeSpec

The new virtual node specification to apply. This overwrites the existing data.

virtualNodeName :: Text

The name of the virtual node to update.

newUpdateVirtualNode

Arguments

:: Text
$sel:meshName:UpdateVirtualNode'

-> VirtualNodeSpec
$sel:spec:UpdateVirtualNode'

-> Text
$sel:virtualNodeName:UpdateVirtualNode'

-> UpdateVirtualNode

Create a value of UpdateVirtualNode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateVirtualNode', updateVirtualNode_clientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

$sel:meshOwner:UpdateVirtualNode', updateVirtualNode_meshOwner - The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

$sel:meshName:UpdateVirtualNode', updateVirtualNode_meshName - The name of the service mesh that the virtual node resides in.

$sel:spec:UpdateVirtualNode', updateVirtualNode_spec - The new virtual node specification to apply. This overwrites the existing data.

$sel:virtualNodeName:UpdateVirtualNode', updateVirtualNode_virtualNodeName - The name of the virtual node to update.

Request Lenses

updateVirtualNode_clientToken :: Lens' UpdateVirtualNode (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Up to 36 letters, numbers, hyphens, and underscores are allowed.

updateVirtualNode_meshOwner :: Lens' UpdateVirtualNode (Maybe Text)

The AWS IAM account ID of the service mesh owner. If the account ID is not your own, then it's the ID of the account that shared the mesh with your account. For more information about mesh sharing, see Working with shared meshes.

updateVirtualNode_meshName :: Lens' UpdateVirtualNode Text

The name of the service mesh that the virtual node resides in.

updateVirtualNode_spec :: Lens' UpdateVirtualNode VirtualNodeSpec

The new virtual node specification to apply. This overwrites the existing data.

updateVirtualNode_virtualNodeName :: Lens' UpdateVirtualNode Text

The name of the virtual node to update.

Destructuring the Response

data UpdateVirtualNodeResponse

See: newUpdateVirtualNodeResponse smart constructor.

Constructors

UpdateVirtualNodeResponse'

Fields

httpStatus :: Int

The response's http status code.

virtualNode :: VirtualNodeData

A full description of the virtual node that was updated.

newUpdateVirtualNodeResponse

Arguments

:: Int
$sel:httpStatus:UpdateVirtualNodeResponse'

-> VirtualNodeData
$sel:virtualNode:UpdateVirtualNodeResponse'

-> UpdateVirtualNodeResponse

Create a value of UpdateVirtualNodeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateVirtualNodeResponse', updateVirtualNodeResponse_httpStatus - The response's http status code.

$sel:virtualNode:UpdateVirtualNodeResponse', updateVirtualNodeResponse_virtualNode - A full description of the virtual node that was updated.

Response Lenses

updateVirtualNodeResponse_httpStatus :: Lens' UpdateVirtualNodeResponse Int

The response's http status code.

updateVirtualNodeResponse_virtualNode :: Lens' UpdateVirtualNodeResponse VirtualNodeData

A full description of the virtual node that was updated.