Network.AWS.AppRunner.DeleteAutoScalingConfiguration

Description

Delete an App Runner automatic scaling configuration resource. You can delete a specific revision or the latest active revision. You can't delete a configuration that's used by one or more App Runner services.

Creating a Request

data DeleteAutoScalingConfiguration

See: newDeleteAutoScalingConfiguration smart constructor.

Constructors

DeleteAutoScalingConfiguration'

Fields

autoScalingConfigurationArn :: Text

The Amazon Resource Name (ARN) of the App Runner auto scaling configuration that you want to delete.

The ARN can be a full auto scaling configuration ARN, or a partial ARN ending with either .../name or .../name/revision . If a revision isn't specified, the latest active revision is deleted.

newDeleteAutoScalingConfiguration

Arguments

:: Text
$sel:autoScalingConfigurationArn:DeleteAutoScalingConfiguration'

-> DeleteAutoScalingConfiguration

Create a value of DeleteAutoScalingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoScalingConfigurationArn:DeleteAutoScalingConfiguration', deleteAutoScalingConfiguration_autoScalingConfigurationArn - The Amazon Resource Name (ARN) of the App Runner auto scaling configuration that you want to delete.

The ARN can be a full auto scaling configuration ARN, or a partial ARN ending with either .../name or .../name/revision . If a revision isn't specified, the latest active revision is deleted.

Request Lenses

deleteAutoScalingConfiguration_autoScalingConfigurationArn :: Lens' DeleteAutoScalingConfiguration Text

The Amazon Resource Name (ARN) of the App Runner auto scaling configuration that you want to delete.

The ARN can be a full auto scaling configuration ARN, or a partial ARN ending with either .../name or .../name/revision . If a revision isn't specified, the latest active revision is deleted.

Destructuring the Response

data DeleteAutoScalingConfigurationResponse

See: newDeleteAutoScalingConfigurationResponse smart constructor.

Constructors

DeleteAutoScalingConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

autoScalingConfiguration :: AutoScalingConfiguration

A description of the App Runner auto scaling configuration that this request just deleted.

newDeleteAutoScalingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DeleteAutoScalingConfigurationResponse'

-> AutoScalingConfiguration
$sel:autoScalingConfiguration:DeleteAutoScalingConfigurationResponse'

-> DeleteAutoScalingConfigurationResponse

Create a value of DeleteAutoScalingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteAutoScalingConfigurationResponse', deleteAutoScalingConfigurationResponse_httpStatus - The response's http status code.

$sel:autoScalingConfiguration:DeleteAutoScalingConfigurationResponse', deleteAutoScalingConfigurationResponse_autoScalingConfiguration - A description of the App Runner auto scaling configuration that this request just deleted.

Response Lenses

deleteAutoScalingConfigurationResponse_httpStatus :: Lens' DeleteAutoScalingConfigurationResponse Int

The response's http status code.

deleteAutoScalingConfigurationResponse_autoScalingConfiguration :: Lens' DeleteAutoScalingConfigurationResponse AutoScalingConfiguration

A description of the App Runner auto scaling configuration that this request just deleted.