Network.AWS.AppRunner.DisassociateCustomDomain

Description

Disassociate a custom domain name from an App Runner service.

Certificates tracking domain validity are associated with a custom domain and are stored in AWS Certificate Manager (ACM). These certificates aren't deleted as part of this action. App Runner delays certificate deletion for 30 days after a domain is disassociated from your service.

Creating a Request

data DisassociateCustomDomain

See: newDisassociateCustomDomain smart constructor.

Constructors

DisassociateCustomDomain'

Fields

serviceArn :: Text

The Amazon Resource Name (ARN) of the App Runner service that you want to disassociate a custom domain name from.

domainName :: Text

The domain name that you want to disassociate from the App Runner service.

newDisassociateCustomDomain

Arguments

:: Text
$sel:serviceArn:DisassociateCustomDomain'

-> Text
$sel:domainName:DisassociateCustomDomain'

-> DisassociateCustomDomain

Create a value of DisassociateCustomDomain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceArn:DisassociateCustomDomain', disassociateCustomDomain_serviceArn - The Amazon Resource Name (ARN) of the App Runner service that you want to disassociate a custom domain name from.

$sel:domainName:DisassociateCustomDomain', disassociateCustomDomain_domainName - The domain name that you want to disassociate from the App Runner service.

Request Lenses

disassociateCustomDomain_serviceArn :: Lens' DisassociateCustomDomain Text

The Amazon Resource Name (ARN) of the App Runner service that you want to disassociate a custom domain name from.

disassociateCustomDomain_domainName :: Lens' DisassociateCustomDomain Text

The domain name that you want to disassociate from the App Runner service.

Destructuring the Response

data DisassociateCustomDomainResponse

See: newDisassociateCustomDomainResponse smart constructor.

Constructors

DisassociateCustomDomainResponse'

Fields

httpStatus :: Int

The response's http status code.

dNSTarget :: Text

The App Runner subdomain of the App Runner service. The disassociated custom domain name was mapped to this target name.

serviceArn :: Text

The Amazon Resource Name (ARN) of the App Runner service that a custom domain name is disassociated from.

customDomain :: CustomDomain

A description of the domain name that's being disassociated.

newDisassociateCustomDomainResponse

Arguments

:: Int
$sel:httpStatus:DisassociateCustomDomainResponse'

-> Text
$sel:dNSTarget:DisassociateCustomDomainResponse'

-> Text
$sel:serviceArn:DisassociateCustomDomain'

-> CustomDomain
$sel:customDomain:DisassociateCustomDomainResponse'

-> DisassociateCustomDomainResponse

Create a value of DisassociateCustomDomainResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateCustomDomainResponse', disassociateCustomDomainResponse_httpStatus - The response's http status code.

$sel:dNSTarget:DisassociateCustomDomainResponse', disassociateCustomDomainResponse_dNSTarget - The App Runner subdomain of the App Runner service. The disassociated custom domain name was mapped to this target name.

$sel:serviceArn:DisassociateCustomDomain', disassociateCustomDomainResponse_serviceArn - The Amazon Resource Name (ARN) of the App Runner service that a custom domain name is disassociated from.

$sel:customDomain:DisassociateCustomDomainResponse', disassociateCustomDomainResponse_customDomain - A description of the domain name that's being disassociated.

Response Lenses

disassociateCustomDomainResponse_httpStatus :: Lens' DisassociateCustomDomainResponse Int

The response's http status code.

disassociateCustomDomainResponse_dNSTarget :: Lens' DisassociateCustomDomainResponse Text

The App Runner subdomain of the App Runner service. The disassociated custom domain name was mapped to this target name.

disassociateCustomDomainResponse_serviceArn :: Lens' DisassociateCustomDomainResponse Text

The Amazon Resource Name (ARN) of the App Runner service that a custom domain name is disassociated from.

disassociateCustomDomainResponse_customDomain :: Lens' DisassociateCustomDomainResponse CustomDomain

A description of the domain name that's being disassociated.