Network.AWS.AppRunner.Types.ServiceSummary

Description

Documentation

data ServiceSummary

Provides summary information for an App Runner service.

This type contains limited information about a service. It doesn't include configuration details. It's returned by the ListServices action. Complete service information is returned by the CreateService, DescribeService, and DeleteService actions using the Service type.

See: newServiceSummary smart constructor.

Constructors

ServiceSummary'

Fields

status :: Maybe ServiceStatus

The current state of the App Runner service. These particular values mean the following.

CREATE_FAILED – The service failed to create. Read the failure events and logs, change any parameters that need to be fixed, and retry the call to create the service.

The failed service isn't usable, and still counts towards your service quota. When you're done analyzing the failure, delete the service.

DELETE_FAILED – The service failed to delete and can't be successfully recovered. Retry the service deletion call to ensure that all related resources are removed.

createdAt :: Maybe POSIX

The time when the App Runner service was created. It's in the Unix time stamp format.

serviceUrl :: Maybe Text

A subdomain URL that App Runner generated for this service. You can use this URL to access your service web application.

serviceName :: Maybe Text

The customer-provided service name.

updatedAt :: Maybe POSIX

The time when the App Runner service was last updated. It's in theUnix time stamp format.

serviceArn :: Maybe Text

The Amazon Resource Name (ARN) of this service.

serviceId :: Maybe Text

An ID that App Runner generated for this service. It's unique within the Amazon Web Services Region.

newServiceSummary :: ServiceSummary

Create a value of ServiceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ServiceSummary', serviceSummary_status - The current state of the App Runner service. These particular values mean the following.

CREATE_FAILED – The service failed to create. Read the failure events and logs, change any parameters that need to be fixed, and retry the call to create the service.

The failed service isn't usable, and still counts towards your service quota. When you're done analyzing the failure, delete the service.

DELETE_FAILED – The service failed to delete and can't be successfully recovered. Retry the service deletion call to ensure that all related resources are removed.

$sel:createdAt:ServiceSummary', serviceSummary_createdAt - The time when the App Runner service was created. It's in the Unix time stamp format.

$sel:serviceUrl:ServiceSummary', serviceSummary_serviceUrl - A subdomain URL that App Runner generated for this service. You can use this URL to access your service web application.

$sel:serviceName:ServiceSummary', serviceSummary_serviceName - The customer-provided service name.

$sel:updatedAt:ServiceSummary', serviceSummary_updatedAt - The time when the App Runner service was last updated. It's in theUnix time stamp format.

$sel:serviceArn:ServiceSummary', serviceSummary_serviceArn - The Amazon Resource Name (ARN) of this service.

$sel:serviceId:ServiceSummary', serviceSummary_serviceId - An ID that App Runner generated for this service. It's unique within the Amazon Web Services Region.

serviceSummary_status :: Lens' ServiceSummary (Maybe ServiceStatus)

The current state of the App Runner service. These particular values mean the following.

CREATE_FAILED – The service failed to create. Read the failure events and logs, change any parameters that need to be fixed, and retry the call to create the service.

The failed service isn't usable, and still counts towards your service quota. When you're done analyzing the failure, delete the service.

DELETE_FAILED – The service failed to delete and can't be successfully recovered. Retry the service deletion call to ensure that all related resources are removed.

serviceSummary_createdAt :: Lens' ServiceSummary (Maybe UTCTime)

The time when the App Runner service was created. It's in the Unix time stamp format.

serviceSummary_serviceUrl :: Lens' ServiceSummary (Maybe Text)

A subdomain URL that App Runner generated for this service. You can use this URL to access your service web application.

serviceSummary_serviceName :: Lens' ServiceSummary (Maybe Text)

The customer-provided service name.

serviceSummary_updatedAt :: Lens' ServiceSummary (Maybe UTCTime)

The time when the App Runner service was last updated. It's in theUnix time stamp format.

serviceSummary_serviceArn :: Lens' ServiceSummary (Maybe Text)

The Amazon Resource Name (ARN) of this service.

serviceSummary_serviceId :: Lens' ServiceSummary (Maybe Text)

An ID that App Runner generated for this service. It's unique within the Amazon Web Services Region.