Network.AWS.AppRunner.Types.SourceConfiguration

Description

Documentation

data SourceConfiguration

Describes the source deployed to an App Runner service. It can be a code or an image repository.

See: newSourceConfiguration smart constructor.

Constructors

SourceConfiguration'

Fields

imageRepository :: Maybe ImageRepository

The description of a source image repository.

You must provide either this member or CodeRepository (but not both).

codeRepository :: Maybe CodeRepository

The description of a source code repository.

You must provide either this member or ImageRepository (but not both).

autoDeploymentsEnabled :: Maybe Bool

If true, continuous integration from the source repository is enabled for the App Runner service. Each repository change (including any source code commit or new image version) starts a deployment.

Default: App Runner sets to false for a source image that uses an ECR Public repository or an ECR repository that's in an Amazon Web Services account other than the one that the service is in. App Runner sets to true in all other cases (which currently include a source code repository or a source image using a same-account ECR repository).

authenticationConfiguration :: Maybe AuthenticationConfiguration

Describes the resources that are needed to authenticate access to some source repositories.

newSourceConfiguration :: SourceConfiguration

Create a value of SourceConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageRepository:SourceConfiguration', sourceConfiguration_imageRepository - The description of a source image repository.

You must provide either this member or CodeRepository (but not both).

$sel:codeRepository:SourceConfiguration', sourceConfiguration_codeRepository - The description of a source code repository.

You must provide either this member or ImageRepository (but not both).

$sel:autoDeploymentsEnabled:SourceConfiguration', sourceConfiguration_autoDeploymentsEnabled - If true, continuous integration from the source repository is enabled for the App Runner service. Each repository change (including any source code commit or new image version) starts a deployment.

Default: App Runner sets to false for a source image that uses an ECR Public repository or an ECR repository that's in an Amazon Web Services account other than the one that the service is in. App Runner sets to true in all other cases (which currently include a source code repository or a source image using a same-account ECR repository).

$sel:authenticationConfiguration:SourceConfiguration', sourceConfiguration_authenticationConfiguration - Describes the resources that are needed to authenticate access to some source repositories.

sourceConfiguration_imageRepository :: Lens' SourceConfiguration (Maybe ImageRepository)

The description of a source image repository.

You must provide either this member or CodeRepository (but not both).

sourceConfiguration_codeRepository :: Lens' SourceConfiguration (Maybe CodeRepository)

The description of a source code repository.

You must provide either this member or ImageRepository (but not both).

sourceConfiguration_autoDeploymentsEnabled :: Lens' SourceConfiguration (Maybe Bool)

If true, continuous integration from the source repository is enabled for the App Runner service. Each repository change (including any source code commit or new image version) starts a deployment.

Default: App Runner sets to false for a source image that uses an ECR Public repository or an ECR repository that's in an Amazon Web Services account other than the one that the service is in. App Runner sets to true in all other cases (which currently include a source code repository or a source image using a same-account ECR repository).

sourceConfiguration_authenticationConfiguration :: Lens' SourceConfiguration (Maybe AuthenticationConfiguration)

Describes the resources that are needed to authenticate access to some source repositories.