Network.AWS.AppStream.CopyImage

Description

Copies the image within the same region or to a new region within the same AWS account. Note that any tags you added to the image will not be copied.

Creating a Request

data CopyImage

See: newCopyImage smart constructor.

Constructors

CopyImage'

Fields

destinationImageDescription :: Maybe Text

The description that the image will have when it is copied to the destination.

sourceImageName :: Text

The name of the image to copy.

destinationImageName :: Text

The name that the image will have when it is copied to the destination.

destinationRegion :: Text

The destination region to which the image will be copied. This parameter is required, even if you are copying an image within the same region.

newCopyImage

Arguments

:: Text
$sel:sourceImageName:CopyImage'

-> Text
$sel:destinationImageName:CopyImage'

-> Text
$sel:destinationRegion:CopyImage'

-> CopyImage

Create a value of CopyImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationImageDescription:CopyImage', copyImage_destinationImageDescription - The description that the image will have when it is copied to the destination.

$sel:sourceImageName:CopyImage', copyImage_sourceImageName - The name of the image to copy.

$sel:destinationImageName:CopyImage', copyImage_destinationImageName - The name that the image will have when it is copied to the destination.

$sel:destinationRegion:CopyImage', copyImage_destinationRegion - The destination region to which the image will be copied. This parameter is required, even if you are copying an image within the same region.

Request Lenses

copyImage_destinationImageDescription :: Lens' CopyImage (Maybe Text)

The description that the image will have when it is copied to the destination.

copyImage_sourceImageName :: Lens' CopyImage Text

The name of the image to copy.

copyImage_destinationImageName :: Lens' CopyImage Text

The name that the image will have when it is copied to the destination.

copyImage_destinationRegion :: Lens' CopyImage Text

The destination region to which the image will be copied. This parameter is required, even if you are copying an image within the same region.

Destructuring the Response

data CopyImageResponse

See: newCopyImageResponse smart constructor.

Constructors

CopyImageResponse'

Fields

destinationImageName :: Maybe Text

The name of the destination image.

httpStatus :: Int

The response's http status code.

newCopyImageResponse

Arguments

:: Int
$sel:httpStatus:CopyImageResponse'

-> CopyImageResponse

Create a value of CopyImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationImageName:CopyImage', copyImageResponse_destinationImageName - The name of the destination image.

$sel:httpStatus:CopyImageResponse', copyImageResponse_httpStatus - The response's http status code.

Response Lenses

copyImageResponse_destinationImageName :: Lens' CopyImageResponse (Maybe Text)

The name of the destination image.

copyImageResponse_httpStatus :: Lens' CopyImageResponse Int

The response's http status code.