Network.AWS.AppStream.CreateDirectoryConfig

Description

Creates a Directory Config object in AppStream 2.0. This object includes the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

Creating a Request

data CreateDirectoryConfig

See: newCreateDirectoryConfig smart constructor.

Constructors

CreateDirectoryConfig'

Fields

serviceAccountCredentials :: Maybe ServiceAccountCredentials

The credentials for the service account used by the fleet or image builder to connect to the directory.

directoryName :: Text

The fully qualified name of the directory (for example, corp.example.com).

organizationalUnitDistinguishedNames :: [Text]

The distinguished names of the organizational units for computer accounts.

newCreateDirectoryConfig

Arguments

:: Text
$sel:directoryName:CreateDirectoryConfig'

-> CreateDirectoryConfig

Create a value of CreateDirectoryConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceAccountCredentials:CreateDirectoryConfig', createDirectoryConfig_serviceAccountCredentials - The credentials for the service account used by the fleet or image builder to connect to the directory.

$sel:directoryName:CreateDirectoryConfig', createDirectoryConfig_directoryName - The fully qualified name of the directory (for example, corp.example.com).

$sel:organizationalUnitDistinguishedNames:CreateDirectoryConfig', createDirectoryConfig_organizationalUnitDistinguishedNames - The distinguished names of the organizational units for computer accounts.

Request Lenses

createDirectoryConfig_serviceAccountCredentials :: Lens' CreateDirectoryConfig (Maybe ServiceAccountCredentials)

The credentials for the service account used by the fleet or image builder to connect to the directory.

createDirectoryConfig_directoryName :: Lens' CreateDirectoryConfig Text

The fully qualified name of the directory (for example, corp.example.com).

createDirectoryConfig_organizationalUnitDistinguishedNames :: Lens' CreateDirectoryConfig [Text]

The distinguished names of the organizational units for computer accounts.

Destructuring the Response

data CreateDirectoryConfigResponse

See: newCreateDirectoryConfigResponse smart constructor.

Constructors

CreateDirectoryConfigResponse'

Fields

directoryConfig :: Maybe DirectoryConfig

Information about the directory configuration.

httpStatus :: Int

The response's http status code.

newCreateDirectoryConfigResponse

Arguments

:: Int
$sel:httpStatus:CreateDirectoryConfigResponse'

-> CreateDirectoryConfigResponse

Create a value of CreateDirectoryConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryConfig:CreateDirectoryConfigResponse', createDirectoryConfigResponse_directoryConfig - Information about the directory configuration.

$sel:httpStatus:CreateDirectoryConfigResponse', createDirectoryConfigResponse_httpStatus - The response's http status code.

Response Lenses

createDirectoryConfigResponse_directoryConfig :: Lens' CreateDirectoryConfigResponse (Maybe DirectoryConfig)

Information about the directory configuration.

createDirectoryConfigResponse_httpStatus :: Lens' CreateDirectoryConfigResponse Int

The response's http status code.