Network.AWS.AppStream.CreateFleet

Description

Creates a fleet. A fleet consists of streaming instances that run a specified image.

Creating a Request

data CreateFleet

See: newCreateFleet smart constructor.

Constructors

CreateFleet'

Fields

domainJoinInfo :: Maybe DomainJoinInfo

The name of the directory and organizational unit (OU) to use to join the fleet to a Microsoft Active Directory domain.

iamRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role to apply to the fleet. To assume a role, a fleet instance calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

disconnectTimeoutInSeconds :: Maybe Int

The amount of time that a streaming session remains active after users disconnect. If users try to reconnect to the streaming session after a disconnection or network interruption within this time interval, they are connected to their previous session. Otherwise, they are connected to a new session with a new streaming instance.

Specify a value between 60 and 360000.

maxUserDurationInSeconds :: Maybe Int

The maximum amount of time that a streaming session can remain active, in seconds. If users are still connected to a streaming instance five minutes before this limit is reached, they are prompted to save any open documents before being disconnected. After this time elapses, the instance is terminated and replaced by a new instance.

Specify a value between 600 and 360000.

idleDisconnectTimeoutInSeconds :: Maybe Int

The amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the DisconnectTimeoutInSeconds time interval begins. Users are notified before they are disconnected due to inactivity. If they try to reconnect to the streaming session before the time interval specified in DisconnectTimeoutInSeconds elapses, they are connected to their previous session. Users are considered idle when they stop providing keyboard or mouse input during their streaming session. File uploads and downloads, audio in, audio out, and pixels changing do not qualify as user activity. If users continue to be idle after the time interval in IdleDisconnectTimeoutInSeconds elapses, they are disconnected.

To prevent users from being disconnected due to inactivity, specify a value of 0. Otherwise, specify a value between 60 and 3600. The default value is 0.

If you enable this feature, we recommend that you specify a value that corresponds exactly to a whole number of minutes (for example, 60, 120, and 180). If you don't do this, the value is rounded to the nearest minute. For example, if you specify a value of 70, users are disconnected after 1 minute of inactivity. If you specify a value that is at the midpoint between two different minutes, the value is rounded up. For example, if you specify a value of 90, users are disconnected after 2 minutes of inactivity.

fleetType :: Maybe FleetType

The fleet type.

ALWAYS_ON

Provides users with instant-on access to their apps. You are charged for all running instances in your fleet, even if no users are streaming apps.

ON_DEMAND

Provide users with access to applications after they connect, which takes one to two minutes. You are charged for instance streaming when users are connected and a small hourly fee for instances that are not streaming apps.

vpcConfig :: Maybe VpcConfig

The VPC configuration for the fleet.

imageArn :: Maybe Text

The ARN of the public, private, or shared image to use.

displayName :: Maybe Text

The fleet name to display.

enableDefaultInternetAccess :: Maybe Bool

Enables or disables default internet access for the fleet.

imageName :: Maybe Text

The name of the image used to create the fleet.

description :: Maybe Text

The description to display.

streamView :: Maybe StreamView

The AppStream 2.0 view that is displayed to your users when they stream from the fleet. When APP is specified, only the windows of applications opened by users display. When DESKTOP is specified, the standard desktop that is provided by the operating system displays.

The default value is APP.

tags :: Maybe (HashMap Text Text)

The tags to associate with the fleet. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

If you do not specify a value, the value is set to an empty string.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

For more information, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

name :: Text

A unique name for the fleet.

instanceType :: Text

The instance type to use when launching fleet instances. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

computeCapacity :: ComputeCapacity

The desired capacity for the fleet.

newCreateFleet

Arguments

:: Text
$sel:name:CreateFleet'

-> Text
$sel:instanceType:CreateFleet'

-> ComputeCapacity
$sel:computeCapacity:CreateFleet'

-> CreateFleet

Create a value of CreateFleet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainJoinInfo:CreateFleet', createFleet_domainJoinInfo - The name of the directory and organizational unit (OU) to use to join the fleet to a Microsoft Active Directory domain.

$sel:iamRoleArn:CreateFleet', createFleet_iamRoleArn - The Amazon Resource Name (ARN) of the IAM role to apply to the fleet. To assume a role, a fleet instance calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

$sel:disconnectTimeoutInSeconds:CreateFleet', createFleet_disconnectTimeoutInSeconds - The amount of time that a streaming session remains active after users disconnect. If users try to reconnect to the streaming session after a disconnection or network interruption within this time interval, they are connected to their previous session. Otherwise, they are connected to a new session with a new streaming instance.

Specify a value between 60 and 360000.

$sel:maxUserDurationInSeconds:CreateFleet', createFleet_maxUserDurationInSeconds - The maximum amount of time that a streaming session can remain active, in seconds. If users are still connected to a streaming instance five minutes before this limit is reached, they are prompted to save any open documents before being disconnected. After this time elapses, the instance is terminated and replaced by a new instance.

Specify a value between 600 and 360000.

$sel:idleDisconnectTimeoutInSeconds:CreateFleet', createFleet_idleDisconnectTimeoutInSeconds - The amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the DisconnectTimeoutInSeconds time interval begins. Users are notified before they are disconnected due to inactivity. If they try to reconnect to the streaming session before the time interval specified in DisconnectTimeoutInSeconds elapses, they are connected to their previous session. Users are considered idle when they stop providing keyboard or mouse input during their streaming session. File uploads and downloads, audio in, audio out, and pixels changing do not qualify as user activity. If users continue to be idle after the time interval in IdleDisconnectTimeoutInSeconds elapses, they are disconnected.

To prevent users from being disconnected due to inactivity, specify a value of 0. Otherwise, specify a value between 60 and 3600. The default value is 0.

If you enable this feature, we recommend that you specify a value that corresponds exactly to a whole number of minutes (for example, 60, 120, and 180). If you don't do this, the value is rounded to the nearest minute. For example, if you specify a value of 70, users are disconnected after 1 minute of inactivity. If you specify a value that is at the midpoint between two different minutes, the value is rounded up. For example, if you specify a value of 90, users are disconnected after 2 minutes of inactivity.

$sel:fleetType:CreateFleet', createFleet_fleetType - The fleet type.

ALWAYS_ON

Provides users with instant-on access to their apps. You are charged for all running instances in your fleet, even if no users are streaming apps.

ON_DEMAND

Provide users with access to applications after they connect, which takes one to two minutes. You are charged for instance streaming when users are connected and a small hourly fee for instances that are not streaming apps.

$sel:vpcConfig:CreateFleet', createFleet_vpcConfig - The VPC configuration for the fleet.

$sel:imageArn:CreateFleet', createFleet_imageArn - The ARN of the public, private, or shared image to use.

$sel:displayName:CreateFleet', createFleet_displayName - The fleet name to display.

$sel:enableDefaultInternetAccess:CreateFleet', createFleet_enableDefaultInternetAccess - Enables or disables default internet access for the fleet.

$sel:imageName:CreateFleet', createFleet_imageName - The name of the image used to create the fleet.

$sel:description:CreateFleet', createFleet_description - The description to display.

$sel:streamView:CreateFleet', createFleet_streamView - The AppStream 2.0 view that is displayed to your users when they stream from the fleet. When APP is specified, only the windows of applications opened by users display. When DESKTOP is specified, the standard desktop that is provided by the operating system displays.

The default value is APP.

$sel:tags:CreateFleet', createFleet_tags - The tags to associate with the fleet. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

If you do not specify a value, the value is set to an empty string.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

For more information, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

$sel:name:CreateFleet', createFleet_name - A unique name for the fleet.

$sel:instanceType:CreateFleet', createFleet_instanceType - The instance type to use when launching fleet instances. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

$sel:computeCapacity:CreateFleet', createFleet_computeCapacity - The desired capacity for the fleet.

Request Lenses

createFleet_domainJoinInfo :: Lens' CreateFleet (Maybe DomainJoinInfo)

The name of the directory and organizational unit (OU) to use to join the fleet to a Microsoft Active Directory domain.

createFleet_iamRoleArn :: Lens' CreateFleet (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role to apply to the fleet. To assume a role, a fleet instance calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

createFleet_disconnectTimeoutInSeconds :: Lens' CreateFleet (Maybe Int)

The amount of time that a streaming session remains active after users disconnect. If users try to reconnect to the streaming session after a disconnection or network interruption within this time interval, they are connected to their previous session. Otherwise, they are connected to a new session with a new streaming instance.

Specify a value between 60 and 360000.

createFleet_maxUserDurationInSeconds :: Lens' CreateFleet (Maybe Int)

The maximum amount of time that a streaming session can remain active, in seconds. If users are still connected to a streaming instance five minutes before this limit is reached, they are prompted to save any open documents before being disconnected. After this time elapses, the instance is terminated and replaced by a new instance.

Specify a value between 600 and 360000.

createFleet_idleDisconnectTimeoutInSeconds :: Lens' CreateFleet (Maybe Int)

The amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the DisconnectTimeoutInSeconds time interval begins. Users are notified before they are disconnected due to inactivity. If they try to reconnect to the streaming session before the time interval specified in DisconnectTimeoutInSeconds elapses, they are connected to their previous session. Users are considered idle when they stop providing keyboard or mouse input during their streaming session. File uploads and downloads, audio in, audio out, and pixels changing do not qualify as user activity. If users continue to be idle after the time interval in IdleDisconnectTimeoutInSeconds elapses, they are disconnected.

To prevent users from being disconnected due to inactivity, specify a value of 0. Otherwise, specify a value between 60 and 3600. The default value is 0.

If you enable this feature, we recommend that you specify a value that corresponds exactly to a whole number of minutes (for example, 60, 120, and 180). If you don't do this, the value is rounded to the nearest minute. For example, if you specify a value of 70, users are disconnected after 1 minute of inactivity. If you specify a value that is at the midpoint between two different minutes, the value is rounded up. For example, if you specify a value of 90, users are disconnected after 2 minutes of inactivity.

createFleet_fleetType :: Lens' CreateFleet (Maybe FleetType)

The fleet type.

ALWAYS_ON

Provides users with instant-on access to their apps. You are charged for all running instances in your fleet, even if no users are streaming apps.

ON_DEMAND

Provide users with access to applications after they connect, which takes one to two minutes. You are charged for instance streaming when users are connected and a small hourly fee for instances that are not streaming apps.

createFleet_vpcConfig :: Lens' CreateFleet (Maybe VpcConfig)

The VPC configuration for the fleet.

createFleet_imageArn :: Lens' CreateFleet (Maybe Text)

The ARN of the public, private, or shared image to use.

createFleet_displayName :: Lens' CreateFleet (Maybe Text)

The fleet name to display.

createFleet_enableDefaultInternetAccess :: Lens' CreateFleet (Maybe Bool)

Enables or disables default internet access for the fleet.

createFleet_imageName :: Lens' CreateFleet (Maybe Text)

The name of the image used to create the fleet.

createFleet_description :: Lens' CreateFleet (Maybe Text)

The description to display.

createFleet_streamView :: Lens' CreateFleet (Maybe StreamView)

The AppStream 2.0 view that is displayed to your users when they stream from the fleet. When APP is specified, only the windows of applications opened by users display. When DESKTOP is specified, the standard desktop that is provided by the operating system displays.

The default value is APP.

createFleet_tags :: Lens' CreateFleet (Maybe (HashMap Text Text))

The tags to associate with the fleet. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

If you do not specify a value, the value is set to an empty string.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

For more information, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

createFleet_name :: Lens' CreateFleet Text

A unique name for the fleet.

createFleet_instanceType :: Lens' CreateFleet Text

The instance type to use when launching fleet instances. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

createFleet_computeCapacity :: Lens' CreateFleet ComputeCapacity

The desired capacity for the fleet.

Destructuring the Response

data CreateFleetResponse

See: newCreateFleetResponse smart constructor.

Constructors

CreateFleetResponse'

Fields

fleet :: Maybe Fleet

Information about the fleet.

httpStatus :: Int

The response's http status code.

newCreateFleetResponse

Arguments

:: Int
$sel:httpStatus:CreateFleetResponse'

-> CreateFleetResponse

Create a value of CreateFleetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fleet:CreateFleetResponse', createFleetResponse_fleet - Information about the fleet.

$sel:httpStatus:CreateFleetResponse', createFleetResponse_httpStatus - The response's http status code.

Response Lenses

createFleetResponse_fleet :: Lens' CreateFleetResponse (Maybe Fleet)

Information about the fleet.

createFleetResponse_httpStatus :: Lens' CreateFleetResponse Int

The response's http status code.