Network.AWS.AppStream.CreateImageBuilder

Description

Creates an image builder. An image builder is a virtual machine that is used to create an image.

The initial state of the builder is PENDING. When it is ready, the state is RUNNING.

Creating a Request

data CreateImageBuilder

See: newCreateImageBuilder smart constructor.

Constructors

CreateImageBuilder'

Fields

vpcConfig :: Maybe VpcConfig

The VPC configuration for the image builder. You can specify only one subnet.

iamRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role to apply to the image builder. To assume a role, the image builder calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

accessEndpoints :: Maybe (NonEmpty AccessEndpoint)

The list of interface VPC endpoint (interface endpoint) objects. Administrators can connect to the image builder only through the specified endpoints.

domainJoinInfo :: Maybe DomainJoinInfo

The name of the directory and organizational unit (OU) to use to join the image builder to a Microsoft Active Directory domain.

imageName :: Maybe Text

The name of the image used to create the image builder.

tags :: Maybe (HashMap Text Text)

The tags to associate with the image builder. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

If you do not specify a value, the value is set to an empty string.

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

appstreamAgentVersion :: Maybe Text

The version of the AppStream 2.0 agent to use for this image builder. To use the latest version of the AppStream 2.0 agent, specify [LATEST].

description :: Maybe Text

The description to display.

enableDefaultInternetAccess :: Maybe Bool

Enables or disables default internet access for the image builder.

displayName :: Maybe Text

The image builder name to display.

imageArn :: Maybe Text

The ARN of the public, private, or shared image to use.

name :: Text

A unique name for the image builder.

instanceType :: Text

The instance type to use when launching the image builder. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

newCreateImageBuilder

Arguments

:: Text
$sel:name:CreateImageBuilder'

-> Text
$sel:instanceType:CreateImageBuilder'

-> CreateImageBuilder

Create a value of CreateImageBuilder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:CreateImageBuilder', createImageBuilder_vpcConfig - The VPC configuration for the image builder. You can specify only one subnet.

$sel:iamRoleArn:CreateImageBuilder', createImageBuilder_iamRoleArn - The Amazon Resource Name (ARN) of the IAM role to apply to the image builder. To assume a role, the image builder calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

$sel:accessEndpoints:CreateImageBuilder', createImageBuilder_accessEndpoints - The list of interface VPC endpoint (interface endpoint) objects. Administrators can connect to the image builder only through the specified endpoints.

$sel:domainJoinInfo:CreateImageBuilder', createImageBuilder_domainJoinInfo - The name of the directory and organizational unit (OU) to use to join the image builder to a Microsoft Active Directory domain.

$sel:imageName:CreateImageBuilder', createImageBuilder_imageName - The name of the image used to create the image builder.

$sel:tags:CreateImageBuilder', createImageBuilder_tags - The tags to associate with the image builder. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

If you do not specify a value, the value is set to an empty string.

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

$sel:appstreamAgentVersion:CreateImageBuilder', createImageBuilder_appstreamAgentVersion - The version of the AppStream 2.0 agent to use for this image builder. To use the latest version of the AppStream 2.0 agent, specify [LATEST].

$sel:description:CreateImageBuilder', createImageBuilder_description - The description to display.

$sel:enableDefaultInternetAccess:CreateImageBuilder', createImageBuilder_enableDefaultInternetAccess - Enables or disables default internet access for the image builder.

$sel:displayName:CreateImageBuilder', createImageBuilder_displayName - The image builder name to display.

$sel:imageArn:CreateImageBuilder', createImageBuilder_imageArn - The ARN of the public, private, or shared image to use.

$sel:name:CreateImageBuilder', createImageBuilder_name - A unique name for the image builder.

$sel:instanceType:CreateImageBuilder', createImageBuilder_instanceType - The instance type to use when launching the image builder. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

Request Lenses

createImageBuilder_vpcConfig :: Lens' CreateImageBuilder (Maybe VpcConfig)

The VPC configuration for the image builder. You can specify only one subnet.

createImageBuilder_iamRoleArn :: Lens' CreateImageBuilder (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role to apply to the image builder. To assume a role, the image builder calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

createImageBuilder_accessEndpoints :: Lens' CreateImageBuilder (Maybe (NonEmpty AccessEndpoint))

The list of interface VPC endpoint (interface endpoint) objects. Administrators can connect to the image builder only through the specified endpoints.

createImageBuilder_domainJoinInfo :: Lens' CreateImageBuilder (Maybe DomainJoinInfo)

The name of the directory and organizational unit (OU) to use to join the image builder to a Microsoft Active Directory domain.

createImageBuilder_imageName :: Lens' CreateImageBuilder (Maybe Text)

The name of the image used to create the image builder.

createImageBuilder_tags :: Lens' CreateImageBuilder (Maybe (HashMap Text Text))

The tags to associate with the image builder. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

If you do not specify a value, the value is set to an empty string.

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

createImageBuilder_appstreamAgentVersion :: Lens' CreateImageBuilder (Maybe Text)

The version of the AppStream 2.0 agent to use for this image builder. To use the latest version of the AppStream 2.0 agent, specify [LATEST].

createImageBuilder_description :: Lens' CreateImageBuilder (Maybe Text)

The description to display.

createImageBuilder_enableDefaultInternetAccess :: Lens' CreateImageBuilder (Maybe Bool)

Enables or disables default internet access for the image builder.

createImageBuilder_displayName :: Lens' CreateImageBuilder (Maybe Text)

The image builder name to display.

createImageBuilder_imageArn :: Lens' CreateImageBuilder (Maybe Text)

The ARN of the public, private, or shared image to use.

createImageBuilder_name :: Lens' CreateImageBuilder Text

A unique name for the image builder.

createImageBuilder_instanceType :: Lens' CreateImageBuilder Text

The instance type to use when launching the image builder. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

Destructuring the Response

data CreateImageBuilderResponse

See: newCreateImageBuilderResponse smart constructor.

Constructors

CreateImageBuilderResponse'

Fields

imageBuilder :: Maybe ImageBuilder

Information about the image builder.

httpStatus :: Int

The response's http status code.

newCreateImageBuilderResponse

Arguments

:: Int
$sel:httpStatus:CreateImageBuilderResponse'

-> CreateImageBuilderResponse

Create a value of CreateImageBuilderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageBuilder:CreateImageBuilderResponse', createImageBuilderResponse_imageBuilder - Information about the image builder.

$sel:httpStatus:CreateImageBuilderResponse', createImageBuilderResponse_httpStatus - The response's http status code.

Response Lenses

createImageBuilderResponse_imageBuilder :: Lens' CreateImageBuilderResponse (Maybe ImageBuilder)

Information about the image builder.

createImageBuilderResponse_httpStatus :: Lens' CreateImageBuilderResponse Int

The response's http status code.