Network.AWS.AppStream.CreateStack

Description

Creates a stack to start streaming applications to users. A stack consists of an associated fleet, user access policies, and storage configurations.

Creating a Request

data CreateStack

See: newCreateStack smart constructor.

Constructors

CreateStack'

Fields

userSettings :: Maybe (NonEmpty UserSetting)

The actions that are enabled or disabled for users during their streaming sessions. By default, these actions are enabled.

accessEndpoints :: Maybe (NonEmpty AccessEndpoint)

The list of interface VPC endpoint (interface endpoint) objects. Users of the stack can connect to AppStream 2.0 only through the specified endpoints.

redirectURL :: Maybe Text

The URL that users are redirected to after their streaming session ends.

applicationSettings :: Maybe ApplicationSettings

The persistent application settings for users of a stack. When these settings are enabled, changes that users make to applications and Windows settings are automatically saved after each session and applied to the next session.

tags :: Maybe (HashMap Text Text)

The tags to associate with the stack. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

If you do not specify a value, the value is set to an empty string.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

storageConnectors :: Maybe [StorageConnector]

The storage connectors to enable.

description :: Maybe Text

The description to display.

embedHostDomains :: Maybe (NonEmpty Text)

The domains where AppStream 2.0 streaming sessions can be embedded in an iframe. You must approve the domains that you want to host embedded AppStream 2.0 streaming sessions.

displayName :: Maybe Text

The stack name to display.

feedbackURL :: Maybe Text

The URL that users are redirected to after they click the Send Feedback link. If no URL is specified, no Send Feedback link is displayed.

name :: Text

The name of the stack.

newCreateStack

Arguments

:: Text
$sel:name:CreateStack'

-> CreateStack

Create a value of CreateStack with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userSettings:CreateStack', createStack_userSettings - The actions that are enabled or disabled for users during their streaming sessions. By default, these actions are enabled.

$sel:accessEndpoints:CreateStack', createStack_accessEndpoints - The list of interface VPC endpoint (interface endpoint) objects. Users of the stack can connect to AppStream 2.0 only through the specified endpoints.

$sel:redirectURL:CreateStack', createStack_redirectURL - The URL that users are redirected to after their streaming session ends.

$sel:applicationSettings:CreateStack', createStack_applicationSettings - The persistent application settings for users of a stack. When these settings are enabled, changes that users make to applications and Windows settings are automatically saved after each session and applied to the next session.

$sel:tags:CreateStack', createStack_tags - The tags to associate with the stack. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

If you do not specify a value, the value is set to an empty string.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

$sel:storageConnectors:CreateStack', createStack_storageConnectors - The storage connectors to enable.

$sel:description:CreateStack', createStack_description - The description to display.

$sel:embedHostDomains:CreateStack', createStack_embedHostDomains - The domains where AppStream 2.0 streaming sessions can be embedded in an iframe. You must approve the domains that you want to host embedded AppStream 2.0 streaming sessions.

$sel:displayName:CreateStack', createStack_displayName - The stack name to display.

$sel:feedbackURL:CreateStack', createStack_feedbackURL - The URL that users are redirected to after they click the Send Feedback link. If no URL is specified, no Send Feedback link is displayed.

$sel:name:CreateStack', createStack_name - The name of the stack.

Request Lenses

createStack_userSettings :: Lens' CreateStack (Maybe (NonEmpty UserSetting))

The actions that are enabled or disabled for users during their streaming sessions. By default, these actions are enabled.

createStack_accessEndpoints :: Lens' CreateStack (Maybe (NonEmpty AccessEndpoint))

The list of interface VPC endpoint (interface endpoint) objects. Users of the stack can connect to AppStream 2.0 only through the specified endpoints.

createStack_redirectURL :: Lens' CreateStack (Maybe Text)

The URL that users are redirected to after their streaming session ends.

createStack_applicationSettings :: Lens' CreateStack (Maybe ApplicationSettings)

The persistent application settings for users of a stack. When these settings are enabled, changes that users make to applications and Windows settings are automatically saved after each session and applied to the next session.

createStack_tags :: Lens' CreateStack (Maybe (HashMap Text Text))

The tags to associate with the stack. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

If you do not specify a value, the value is set to an empty string.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

createStack_storageConnectors :: Lens' CreateStack (Maybe [StorageConnector])

The storage connectors to enable.

createStack_description :: Lens' CreateStack (Maybe Text)

The description to display.

createStack_embedHostDomains :: Lens' CreateStack (Maybe (NonEmpty Text))

The domains where AppStream 2.0 streaming sessions can be embedded in an iframe. You must approve the domains that you want to host embedded AppStream 2.0 streaming sessions.

createStack_displayName :: Lens' CreateStack (Maybe Text)

The stack name to display.

createStack_feedbackURL :: Lens' CreateStack (Maybe Text)

The URL that users are redirected to after they click the Send Feedback link. If no URL is specified, no Send Feedback link is displayed.

createStack_name :: Lens' CreateStack Text

The name of the stack.

Destructuring the Response

data CreateStackResponse

See: newCreateStackResponse smart constructor.

Constructors

CreateStackResponse'

Fields

stack :: Maybe Stack

Information about the stack.

httpStatus :: Int

The response's http status code.

newCreateStackResponse

Arguments

:: Int
$sel:httpStatus:CreateStackResponse'

-> CreateStackResponse

Create a value of CreateStackResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stack:CreateStackResponse', createStackResponse_stack - Information about the stack.

$sel:httpStatus:CreateStackResponse', createStackResponse_httpStatus - The response's http status code.

Response Lenses

createStackResponse_stack :: Lens' CreateStackResponse (Maybe Stack)

Information about the stack.

createStackResponse_httpStatus :: Lens' CreateStackResponse Int

The response's http status code.