Network.AWS.AppStream.CreateStreamingURL

Description

Creates a temporary URL to start an AppStream 2.0 streaming session for the specified user. A streaming URL enables application streaming to be tested without user setup.

Creating a Request

data CreateStreamingURL

See: newCreateStreamingURL smart constructor.

Constructors

CreateStreamingURL'

Fields

applicationId :: Maybe Text

The name of the application to launch after the session starts. This is the name that you specified as Name in the Image Assistant. If your fleet is enabled for the Desktop stream view, you can also choose to launch directly to the operating system desktop. To do so, specify Desktop.

sessionContext :: Maybe Text

The session context. For more information, see Session Context in the Amazon AppStream 2.0 Administration Guide.

validity :: Maybe Integer

The time that the streaming URL will be valid, in seconds. Specify a value between 1 and 604800 seconds. The default is 60 seconds.

stackName :: Text

The name of the stack.

fleetName :: Text

The name of the fleet.

userId :: Text

The identifier of the user.

newCreateStreamingURL

Arguments

:: Text
$sel:stackName:CreateStreamingURL'

-> Text
$sel:fleetName:CreateStreamingURL'

-> Text
$sel:userId:CreateStreamingURL'

-> CreateStreamingURL

Create a value of CreateStreamingURL with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:CreateStreamingURL', createStreamingURL_applicationId - The name of the application to launch after the session starts. This is the name that you specified as Name in the Image Assistant. If your fleet is enabled for the Desktop stream view, you can also choose to launch directly to the operating system desktop. To do so, specify Desktop.

$sel:sessionContext:CreateStreamingURL', createStreamingURL_sessionContext - The session context. For more information, see Session Context in the Amazon AppStream 2.0 Administration Guide.

$sel:validity:CreateStreamingURL', createStreamingURL_validity - The time that the streaming URL will be valid, in seconds. Specify a value between 1 and 604800 seconds. The default is 60 seconds.

$sel:stackName:CreateStreamingURL', createStreamingURL_stackName - The name of the stack.

$sel:fleetName:CreateStreamingURL', createStreamingURL_fleetName - The name of the fleet.

$sel:userId:CreateStreamingURL', createStreamingURL_userId - The identifier of the user.

Request Lenses

createStreamingURL_applicationId :: Lens' CreateStreamingURL (Maybe Text)

The name of the application to launch after the session starts. This is the name that you specified as Name in the Image Assistant. If your fleet is enabled for the Desktop stream view, you can also choose to launch directly to the operating system desktop. To do so, specify Desktop.

createStreamingURL_sessionContext :: Lens' CreateStreamingURL (Maybe Text)

The session context. For more information, see Session Context in the Amazon AppStream 2.0 Administration Guide.

createStreamingURL_validity :: Lens' CreateStreamingURL (Maybe Integer)

The time that the streaming URL will be valid, in seconds. Specify a value between 1 and 604800 seconds. The default is 60 seconds.

createStreamingURL_stackName :: Lens' CreateStreamingURL Text

The name of the stack.

createStreamingURL_fleetName :: Lens' CreateStreamingURL Text

The name of the fleet.

createStreamingURL_userId :: Lens' CreateStreamingURL Text

The identifier of the user.

Destructuring the Response

data CreateStreamingURLResponse

See: newCreateStreamingURLResponse smart constructor.

Constructors

CreateStreamingURLResponse'

Fields

streamingURL :: Maybe Text

The URL to start the AppStream 2.0 streaming session.

expires :: Maybe POSIX

The elapsed time, in seconds after the Unix epoch, when this URL expires.

httpStatus :: Int

The response's http status code.

newCreateStreamingURLResponse

Arguments

:: Int
$sel:httpStatus:CreateStreamingURLResponse'

-> CreateStreamingURLResponse

Create a value of CreateStreamingURLResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamingURL:CreateStreamingURLResponse', createStreamingURLResponse_streamingURL - The URL to start the AppStream 2.0 streaming session.

$sel:expires:CreateStreamingURLResponse', createStreamingURLResponse_expires - The elapsed time, in seconds after the Unix epoch, when this URL expires.

$sel:httpStatus:CreateStreamingURLResponse', createStreamingURLResponse_httpStatus - The response's http status code.

Response Lenses

createStreamingURLResponse_streamingURL :: Lens' CreateStreamingURLResponse (Maybe Text)

The URL to start the AppStream 2.0 streaming session.

createStreamingURLResponse_expires :: Lens' CreateStreamingURLResponse (Maybe UTCTime)

The elapsed time, in seconds after the Unix epoch, when this URL expires.

createStreamingURLResponse_httpStatus :: Lens' CreateStreamingURLResponse Int

The response's http status code.