Network.AWS.AppStream.CreateUpdatedImage

Description

Creates a new image with the latest Windows operating system updates, driver updates, and AppStream 2.0 agent software.

For more information, see the "Update an Image by Using Managed AppStream 2.0 Image Updates" section in Administer Your AppStream 2.0 Images, in the Amazon AppStream 2.0 Administration Guide.

Creating a Request

data CreateUpdatedImage

See: newCreateUpdatedImage smart constructor.

Constructors

CreateUpdatedImage'

Fields

newImageDisplayName' :: Maybe Text

The name to display for the new image.

dryRun :: Maybe Bool

Indicates whether to display the status of image update availability before AppStream 2.0 initiates the process of creating a new updated image. If this value is set to true, AppStream 2.0 displays whether image updates are available. If this value is set to false, AppStream 2.0 initiates the process of creating a new updated image without displaying whether image updates are available.

newImageDescription' :: Maybe Text

The description to display for the new image.

newImageTags' :: Maybe (HashMap Text Text)

The tags to associate with the new image. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

If you do not specify a value, the value is set to an empty string.

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

existingImageName :: Text

The name of the image to update.

newImageName' :: Text

The name of the new image. The name must be unique within the AWS account and Region.

newCreateUpdatedImage

Arguments

:: Text
$sel:existingImageName:CreateUpdatedImage'

-> Text
$sel:newImageName':CreateUpdatedImage'

-> CreateUpdatedImage

Create a value of CreateUpdatedImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:newImageDisplayName':CreateUpdatedImage', createUpdatedImage_newImageDisplayName - The name to display for the new image.

$sel:dryRun:CreateUpdatedImage', createUpdatedImage_dryRun - Indicates whether to display the status of image update availability before AppStream 2.0 initiates the process of creating a new updated image. If this value is set to true, AppStream 2.0 displays whether image updates are available. If this value is set to false, AppStream 2.0 initiates the process of creating a new updated image without displaying whether image updates are available.

$sel:newImageDescription':CreateUpdatedImage', createUpdatedImage_newImageDescription - The description to display for the new image.

$sel:newImageTags':CreateUpdatedImage', createUpdatedImage_newImageTags - The tags to associate with the new image. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

If you do not specify a value, the value is set to an empty string.

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

$sel:existingImageName:CreateUpdatedImage', createUpdatedImage_existingImageName - The name of the image to update.

$sel:newImageName':CreateUpdatedImage', createUpdatedImage_newImageName - The name of the new image. The name must be unique within the AWS account and Region.

Request Lenses

createUpdatedImage_newImageDisplayName :: Lens' CreateUpdatedImage (Maybe Text)

The name to display for the new image.

createUpdatedImage_dryRun :: Lens' CreateUpdatedImage (Maybe Bool)

Indicates whether to display the status of image update availability before AppStream 2.0 initiates the process of creating a new updated image. If this value is set to true, AppStream 2.0 displays whether image updates are available. If this value is set to false, AppStream 2.0 initiates the process of creating a new updated image without displaying whether image updates are available.

createUpdatedImage_newImageDescription :: Lens' CreateUpdatedImage (Maybe Text)

The description to display for the new image.

createUpdatedImage_newImageTags :: Lens' CreateUpdatedImage (Maybe (HashMap Text Text))

The tags to associate with the new image. A tag is a key-value pair, and the value is optional. For example, Environment=Test. If you do not specify a value, Environment=.

Generally allowed characters are: letters, numbers, and spaces representable in UTF-8, and the following special characters:

_ . : / = + \ - @

If you do not specify a value, the value is set to an empty string.

For more information about tags, see Tagging Your Resources in the Amazon AppStream 2.0 Administration Guide.

createUpdatedImage_existingImageName :: Lens' CreateUpdatedImage Text

The name of the image to update.

createUpdatedImage_newImageName :: Lens' CreateUpdatedImage Text

The name of the new image. The name must be unique within the AWS account and Region.

Destructuring the Response

data CreateUpdatedImageResponse

See: newCreateUpdatedImageResponse smart constructor.

Constructors

CreateUpdatedImageResponse'

Fields

image :: Maybe Image

canUpdateImage :: Maybe Bool

Indicates whether a new image can be created.

httpStatus :: Int

The response's http status code.

newCreateUpdatedImageResponse

Arguments

:: Int
$sel:httpStatus:CreateUpdatedImageResponse'

-> CreateUpdatedImageResponse

Create a value of CreateUpdatedImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:image:CreateUpdatedImageResponse', createUpdatedImageResponse_image - Undocumented member.

$sel:canUpdateImage:CreateUpdatedImageResponse', createUpdatedImageResponse_canUpdateImage - Indicates whether a new image can be created.

$sel:httpStatus:CreateUpdatedImageResponse', createUpdatedImageResponse_httpStatus - The response's http status code.

Response Lenses

createUpdatedImageResponse_image :: Lens' CreateUpdatedImageResponse (Maybe Image)

Undocumented member.

createUpdatedImageResponse_canUpdateImage :: Lens' CreateUpdatedImageResponse (Maybe Bool)

Indicates whether a new image can be created.

createUpdatedImageResponse_httpStatus :: Lens' CreateUpdatedImageResponse Int

The response's http status code.