Network.AWS.AppStream.DeleteImagePermissions

Description

Deletes permissions for the specified private image. After you delete permissions for an image, AWS accounts to which you previously granted these permissions can no longer use the image.

Creating a Request

data DeleteImagePermissions

See: newDeleteImagePermissions smart constructor.

Constructors

DeleteImagePermissions'

Fields

name :: Text

The name of the private image.

sharedAccountId :: Text

The 12-digit identifier of the AWS account for which to delete image permissions.

newDeleteImagePermissions

Arguments

:: Text
$sel:name:DeleteImagePermissions'

-> Text
$sel:sharedAccountId:DeleteImagePermissions'

-> DeleteImagePermissions

Create a value of DeleteImagePermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteImagePermissions', deleteImagePermissions_name - The name of the private image.

$sel:sharedAccountId:DeleteImagePermissions', deleteImagePermissions_sharedAccountId - The 12-digit identifier of the AWS account for which to delete image permissions.

Request Lenses

deleteImagePermissions_name :: Lens' DeleteImagePermissions Text

The name of the private image.

deleteImagePermissions_sharedAccountId :: Lens' DeleteImagePermissions Text

The 12-digit identifier of the AWS account for which to delete image permissions.

Destructuring the Response

data DeleteImagePermissionsResponse

See: newDeleteImagePermissionsResponse smart constructor.

Constructors

DeleteImagePermissionsResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteImagePermissionsResponse

Arguments

:: Int
$sel:httpStatus:DeleteImagePermissionsResponse'

-> DeleteImagePermissionsResponse

Create a value of DeleteImagePermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteImagePermissionsResponse', deleteImagePermissionsResponse_httpStatus - The response's http status code.

Response Lenses

deleteImagePermissionsResponse_httpStatus :: Lens' DeleteImagePermissionsResponse Int

The response's http status code.