Network.AWS.AppStream.DescribeDirectoryConfigs

Description

Retrieves a list that describes one or more specified Directory Config objects for AppStream 2.0, if the names for these objects are provided. Otherwise, all Directory Config objects in the account are described. These objects include the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

Although the response syntax in this topic includes the account password, this password is not returned in the actual response.

This operation returns paginated results.

Creating a Request

data DescribeDirectoryConfigs

See: newDescribeDirectoryConfigs smart constructor.

Constructors

DescribeDirectoryConfigs'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

maxResults :: Maybe Int

The maximum size of each page of results.

directoryNames :: Maybe [Text]

The directory names.

newDescribeDirectoryConfigs :: DescribeDirectoryConfigs

Create a value of DescribeDirectoryConfigs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDirectoryConfigs', describeDirectoryConfigs_nextToken - The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

$sel:maxResults:DescribeDirectoryConfigs', describeDirectoryConfigs_maxResults - The maximum size of each page of results.

$sel:directoryNames:DescribeDirectoryConfigs', describeDirectoryConfigs_directoryNames - The directory names.

Request Lenses

describeDirectoryConfigs_nextToken :: Lens' DescribeDirectoryConfigs (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

describeDirectoryConfigs_maxResults :: Lens' DescribeDirectoryConfigs (Maybe Int)

The maximum size of each page of results.

describeDirectoryConfigs_directoryNames :: Lens' DescribeDirectoryConfigs (Maybe [Text])

The directory names.

Destructuring the Response

data DescribeDirectoryConfigsResponse

See: newDescribeDirectoryConfigsResponse smart constructor.

Constructors

DescribeDirectoryConfigsResponse'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

directoryConfigs :: Maybe [DirectoryConfig]

Information about the directory configurations. Note that although the response syntax in this topic includes the account password, this password is not returned in the actual response.

httpStatus :: Int

The response's http status code.

newDescribeDirectoryConfigsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDirectoryConfigsResponse'

-> DescribeDirectoryConfigsResponse

Create a value of DescribeDirectoryConfigsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDirectoryConfigs', describeDirectoryConfigsResponse_nextToken - The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

$sel:directoryConfigs:DescribeDirectoryConfigsResponse', describeDirectoryConfigsResponse_directoryConfigs - Information about the directory configurations. Note that although the response syntax in this topic includes the account password, this password is not returned in the actual response.

$sel:httpStatus:DescribeDirectoryConfigsResponse', describeDirectoryConfigsResponse_httpStatus - The response's http status code.

Response Lenses

describeDirectoryConfigsResponse_nextToken :: Lens' DescribeDirectoryConfigsResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

describeDirectoryConfigsResponse_directoryConfigs :: Lens' DescribeDirectoryConfigsResponse (Maybe [DirectoryConfig])

Information about the directory configurations. Note that although the response syntax in this topic includes the account password, this password is not returned in the actual response.

describeDirectoryConfigsResponse_httpStatus :: Lens' DescribeDirectoryConfigsResponse Int

The response's http status code.