Network.AWS.AppStream.DescribeUsageReportSubscriptions

Description

Retrieves a list that describes one or more usage report subscriptions.

Creating a Request

data DescribeUsageReportSubscriptions

See: newDescribeUsageReportSubscriptions smart constructor.

Constructors

DescribeUsageReportSubscriptions'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

maxResults :: Maybe Int

The maximum size of each page of results.

newDescribeUsageReportSubscriptions :: DescribeUsageReportSubscriptions

Create a value of DescribeUsageReportSubscriptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeUsageReportSubscriptions', describeUsageReportSubscriptions_nextToken - The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

$sel:maxResults:DescribeUsageReportSubscriptions', describeUsageReportSubscriptions_maxResults - The maximum size of each page of results.

Request Lenses

describeUsageReportSubscriptions_nextToken :: Lens' DescribeUsageReportSubscriptions (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

describeUsageReportSubscriptions_maxResults :: Lens' DescribeUsageReportSubscriptions (Maybe Int)

The maximum size of each page of results.

Destructuring the Response

data DescribeUsageReportSubscriptionsResponse

See: newDescribeUsageReportSubscriptionsResponse smart constructor.

Constructors

DescribeUsageReportSubscriptionsResponse'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

usageReportSubscriptions :: Maybe [UsageReportSubscription]

Information about the usage report subscription.

httpStatus :: Int

The response's http status code.

newDescribeUsageReportSubscriptionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeUsageReportSubscriptionsResponse'

-> DescribeUsageReportSubscriptionsResponse

Create a value of DescribeUsageReportSubscriptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeUsageReportSubscriptions', describeUsageReportSubscriptionsResponse_nextToken - The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

$sel:usageReportSubscriptions:DescribeUsageReportSubscriptionsResponse', describeUsageReportSubscriptionsResponse_usageReportSubscriptions - Information about the usage report subscription.

$sel:httpStatus:DescribeUsageReportSubscriptionsResponse', describeUsageReportSubscriptionsResponse_httpStatus - The response's http status code.

Response Lenses

describeUsageReportSubscriptionsResponse_nextToken :: Lens' DescribeUsageReportSubscriptionsResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

describeUsageReportSubscriptionsResponse_usageReportSubscriptions :: Lens' DescribeUsageReportSubscriptionsResponse (Maybe [UsageReportSubscription])

Information about the usage report subscription.

describeUsageReportSubscriptionsResponse_httpStatus :: Lens' DescribeUsageReportSubscriptionsResponse Int

The response's http status code.