Network.AWS.AppStream.DescribeUserStackAssociations

Description

Retrieves a list that describes the UserStackAssociation objects. You must specify either or both of the following:

The stack name

The user name (email address of the user associated with the stack) and the authentication type for the user

This operation returns paginated results.

Creating a Request

data DescribeUserStackAssociations

See: newDescribeUserStackAssociations smart constructor.

Constructors

DescribeUserStackAssociations'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

stackName :: Maybe Text

The name of the stack that is associated with the user.

maxResults :: Maybe Natural

The maximum size of each page of results.

userName :: Maybe (Sensitive Text)

The email address of the user who is associated with the stack.

Users' email addresses are case-sensitive.

authenticationType :: Maybe AuthenticationType

The authentication type for the user who is associated with the stack. You must specify USERPOOL.

newDescribeUserStackAssociations :: DescribeUserStackAssociations

Create a value of DescribeUserStackAssociations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeUserStackAssociations', describeUserStackAssociations_nextToken - The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

$sel:stackName:DescribeUserStackAssociations', describeUserStackAssociations_stackName - The name of the stack that is associated with the user.

$sel:maxResults:DescribeUserStackAssociations', describeUserStackAssociations_maxResults - The maximum size of each page of results.

$sel:userName:DescribeUserStackAssociations', describeUserStackAssociations_userName - The email address of the user who is associated with the stack.

Users' email addresses are case-sensitive.

$sel:authenticationType:DescribeUserStackAssociations', describeUserStackAssociations_authenticationType - The authentication type for the user who is associated with the stack. You must specify USERPOOL.

Request Lenses

describeUserStackAssociations_nextToken :: Lens' DescribeUserStackAssociations (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

describeUserStackAssociations_stackName :: Lens' DescribeUserStackAssociations (Maybe Text)

The name of the stack that is associated with the user.

describeUserStackAssociations_maxResults :: Lens' DescribeUserStackAssociations (Maybe Natural)

The maximum size of each page of results.

describeUserStackAssociations_userName :: Lens' DescribeUserStackAssociations (Maybe Text)

The email address of the user who is associated with the stack.

Users' email addresses are case-sensitive.

describeUserStackAssociations_authenticationType :: Lens' DescribeUserStackAssociations (Maybe AuthenticationType)

The authentication type for the user who is associated with the stack. You must specify USERPOOL.

Destructuring the Response

data DescribeUserStackAssociationsResponse

See: newDescribeUserStackAssociationsResponse smart constructor.

Constructors

DescribeUserStackAssociationsResponse'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

userStackAssociations :: Maybe (NonEmpty UserStackAssociation)

The UserStackAssociation objects.

httpStatus :: Int

The response's http status code.

newDescribeUserStackAssociationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeUserStackAssociationsResponse'

-> DescribeUserStackAssociationsResponse

Create a value of DescribeUserStackAssociationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeUserStackAssociations', describeUserStackAssociationsResponse_nextToken - The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

$sel:userStackAssociations:DescribeUserStackAssociationsResponse', describeUserStackAssociationsResponse_userStackAssociations - The UserStackAssociation objects.

$sel:httpStatus:DescribeUserStackAssociationsResponse', describeUserStackAssociationsResponse_httpStatus - The response's http status code.

Response Lenses

describeUserStackAssociationsResponse_nextToken :: Lens' DescribeUserStackAssociationsResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

describeUserStackAssociationsResponse_userStackAssociations :: Lens' DescribeUserStackAssociationsResponse (Maybe (NonEmpty UserStackAssociation))

The UserStackAssociation objects.

describeUserStackAssociationsResponse_httpStatus :: Lens' DescribeUserStackAssociationsResponse Int

The response's http status code.