Network.AWS.AppStream.DescribeUsers

Description

Retrieves a list that describes one or more specified users in the user pool.

This operation returns paginated results.

Creating a Request

data DescribeUsers

See: newDescribeUsers smart constructor.

Constructors

DescribeUsers'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

maxResults :: Maybe Int

The maximum size of each page of results.

authenticationType :: AuthenticationType

The authentication type for the users in the user pool to describe. You must specify USERPOOL.

newDescribeUsers

Arguments

:: AuthenticationType
$sel:authenticationType:DescribeUsers'

-> DescribeUsers

Create a value of DescribeUsers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeUsers', describeUsers_nextToken - The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

$sel:maxResults:DescribeUsers', describeUsers_maxResults - The maximum size of each page of results.

$sel:authenticationType:DescribeUsers', describeUsers_authenticationType - The authentication type for the users in the user pool to describe. You must specify USERPOOL.

Request Lenses

describeUsers_nextToken :: Lens' DescribeUsers (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

describeUsers_maxResults :: Lens' DescribeUsers (Maybe Int)

The maximum size of each page of results.

describeUsers_authenticationType :: Lens' DescribeUsers AuthenticationType

The authentication type for the users in the user pool to describe. You must specify USERPOOL.

Destructuring the Response

data DescribeUsersResponse

See: newDescribeUsersResponse smart constructor.

Constructors

DescribeUsersResponse'

Fields

nextToken :: Maybe Text

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

users :: Maybe [User]

Information about users in the user pool.

httpStatus :: Int

The response's http status code.

newDescribeUsersResponse

Arguments

:: Int
$sel:httpStatus:DescribeUsersResponse'

-> DescribeUsersResponse

Create a value of DescribeUsersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeUsers', describeUsersResponse_nextToken - The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

$sel:users:DescribeUsersResponse', describeUsersResponse_users - Information about users in the user pool.

$sel:httpStatus:DescribeUsersResponse', describeUsersResponse_httpStatus - The response's http status code.

Response Lenses

describeUsersResponse_nextToken :: Lens' DescribeUsersResponse (Maybe Text)

The pagination token to use to retrieve the next page of results for this operation. If there are no more pages, this value is null.

describeUsersResponse_users :: Lens' DescribeUsersResponse (Maybe [User])

Information about users in the user pool.

describeUsersResponse_httpStatus :: Lens' DescribeUsersResponse Int

The response's http status code.