Network.AWS.AppStream.ExpireSession

Description

Immediately stops the specified streaming session.

Creating a Request

data ExpireSession

See: newExpireSession smart constructor.

Constructors

ExpireSession'

Fields

sessionId :: Text

The identifier of the streaming session.

newExpireSession

Arguments

:: Text
$sel:sessionId:ExpireSession'

-> ExpireSession

Create a value of ExpireSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionId:ExpireSession', expireSession_sessionId - The identifier of the streaming session.

Request Lenses

expireSession_sessionId :: Lens' ExpireSession Text

The identifier of the streaming session.

Destructuring the Response

data ExpireSessionResponse

See: newExpireSessionResponse smart constructor.

Constructors

ExpireSessionResponse'

Fields

httpStatus :: Int

The response's http status code.

newExpireSessionResponse

Arguments

:: Int
$sel:httpStatus:ExpireSessionResponse'

-> ExpireSessionResponse

Create a value of ExpireSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ExpireSessionResponse', expireSessionResponse_httpStatus - The response's http status code.

Response Lenses

expireSessionResponse_httpStatus :: Lens' ExpireSessionResponse Int

The response's http status code.