Network.AWS.AppStream.Types.Fleet

Description

Documentation

data Fleet

Describes a fleet.

See: newFleet smart constructor.

Constructors

Fleet'

Fields

disconnectTimeoutInSeconds :: Maybe Int

The amount of time that a streaming session remains active after users disconnect. If they try to reconnect to the streaming session after a disconnection or network interruption within this time interval, they are connected to their previous session. Otherwise, they are connected to a new session with a new streaming instance.

Specify a value between 60 and 360000.

maxUserDurationInSeconds :: Maybe Int

The maximum amount of time that a streaming session can remain active, in seconds. If users are still connected to a streaming instance five minutes before this limit is reached, they are prompted to save any open documents before being disconnected. After this time elapses, the instance is terminated and replaced by a new instance.

Specify a value between 600 and 360000.

vpcConfig :: Maybe VpcConfig

The VPC configuration for the fleet.

iamRoleArn :: Maybe Text

The ARN of the IAM role that is applied to the fleet. To assume a role, the fleet instance calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

domainJoinInfo :: Maybe DomainJoinInfo

The name of the directory and organizational unit (OU) to use to join the fleet to a Microsoft Active Directory domain.

fleetType :: Maybe FleetType

The fleet type.

ALWAYS_ON

Provides users with instant-on access to their apps. You are charged for all running instances in your fleet, even if no users are streaming apps.

ON_DEMAND

Provide users with access to applications after they connect, which takes one to two minutes. You are charged for instance streaming when users are connected and a small hourly fee for instances that are not streaming apps.

idleDisconnectTimeoutInSeconds :: Maybe Int

The amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the DisconnectTimeoutInSeconds time interval begins. Users are notified before they are disconnected due to inactivity. If users try to reconnect to the streaming session before the time interval specified in DisconnectTimeoutInSeconds elapses, they are connected to their previous session. Users are considered idle when they stop providing keyboard or mouse input during their streaming session. File uploads and downloads, audio in, audio out, and pixels changing do not qualify as user activity. If users continue to be idle after the time interval in IdleDisconnectTimeoutInSeconds elapses, they are disconnected.

To prevent users from being disconnected due to inactivity, specify a value of 0. Otherwise, specify a value between 60 and 3600. The default value is 0.

If you enable this feature, we recommend that you specify a value that corresponds exactly to a whole number of minutes (for example, 60, 120, and 180). If you don't do this, the value is rounded to the nearest minute. For example, if you specify a value of 70, users are disconnected after 1 minute of inactivity. If you specify a value that is at the midpoint between two different minutes, the value is rounded up. For example, if you specify a value of 90, users are disconnected after 2 minutes of inactivity.

imageName :: Maybe Text

The name of the image used to create the fleet.

createdTime :: Maybe POSIX

The time the fleet was created.

streamView :: Maybe StreamView

The AppStream 2.0 view that is displayed to your users when they stream from the fleet. When APP is specified, only the windows of applications opened by users display. When DESKTOP is specified, the standard desktop that is provided by the operating system displays.

The default value is APP.

description :: Maybe Text

The description to display.

enableDefaultInternetAccess :: Maybe Bool

Indicates whether default internet access is enabled for the fleet.

fleetErrors :: Maybe [FleetError]

The fleet errors.

displayName :: Maybe Text

The fleet name to display.

imageArn :: Maybe Text

The ARN for the public, private, or shared image.

arn :: Text

The Amazon Resource Name (ARN) for the fleet.

name :: Text

The name of the fleet.

instanceType :: Text

The instance type to use when launching fleet instances. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

computeCapacityStatus :: ComputeCapacityStatus

The capacity status for the fleet.

state :: FleetState

The current state for the fleet.

newFleet

Arguments

:: Text
$sel:arn:Fleet'

-> Text
$sel:name:Fleet'

-> Text
$sel:instanceType:Fleet'

-> ComputeCapacityStatus
$sel:computeCapacityStatus:Fleet'

-> FleetState
$sel:state:Fleet'

-> Fleet

Create a value of Fleet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disconnectTimeoutInSeconds:Fleet', fleet_disconnectTimeoutInSeconds - The amount of time that a streaming session remains active after users disconnect. If they try to reconnect to the streaming session after a disconnection or network interruption within this time interval, they are connected to their previous session. Otherwise, they are connected to a new session with a new streaming instance.

Specify a value between 60 and 360000.

$sel:maxUserDurationInSeconds:Fleet', fleet_maxUserDurationInSeconds - The maximum amount of time that a streaming session can remain active, in seconds. If users are still connected to a streaming instance five minutes before this limit is reached, they are prompted to save any open documents before being disconnected. After this time elapses, the instance is terminated and replaced by a new instance.

Specify a value between 600 and 360000.

$sel:vpcConfig:Fleet', fleet_vpcConfig - The VPC configuration for the fleet.

$sel:iamRoleArn:Fleet', fleet_iamRoleArn - The ARN of the IAM role that is applied to the fleet. To assume a role, the fleet instance calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

$sel:domainJoinInfo:Fleet', fleet_domainJoinInfo - The name of the directory and organizational unit (OU) to use to join the fleet to a Microsoft Active Directory domain.

$sel:fleetType:Fleet', fleet_fleetType - The fleet type.

ALWAYS_ON

Provides users with instant-on access to their apps. You are charged for all running instances in your fleet, even if no users are streaming apps.

ON_DEMAND

Provide users with access to applications after they connect, which takes one to two minutes. You are charged for instance streaming when users are connected and a small hourly fee for instances that are not streaming apps.

$sel:idleDisconnectTimeoutInSeconds:Fleet', fleet_idleDisconnectTimeoutInSeconds - The amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the DisconnectTimeoutInSeconds time interval begins. Users are notified before they are disconnected due to inactivity. If users try to reconnect to the streaming session before the time interval specified in DisconnectTimeoutInSeconds elapses, they are connected to their previous session. Users are considered idle when they stop providing keyboard or mouse input during their streaming session. File uploads and downloads, audio in, audio out, and pixels changing do not qualify as user activity. If users continue to be idle after the time interval in IdleDisconnectTimeoutInSeconds elapses, they are disconnected.

To prevent users from being disconnected due to inactivity, specify a value of 0. Otherwise, specify a value between 60 and 3600. The default value is 0.

If you enable this feature, we recommend that you specify a value that corresponds exactly to a whole number of minutes (for example, 60, 120, and 180). If you don't do this, the value is rounded to the nearest minute. For example, if you specify a value of 70, users are disconnected after 1 minute of inactivity. If you specify a value that is at the midpoint between two different minutes, the value is rounded up. For example, if you specify a value of 90, users are disconnected after 2 minutes of inactivity.

$sel:imageName:Fleet', fleet_imageName - The name of the image used to create the fleet.

$sel:createdTime:Fleet', fleet_createdTime - The time the fleet was created.

$sel:streamView:Fleet', fleet_streamView - The AppStream 2.0 view that is displayed to your users when they stream from the fleet. When APP is specified, only the windows of applications opened by users display. When DESKTOP is specified, the standard desktop that is provided by the operating system displays.

The default value is APP.

$sel:description:Fleet', fleet_description - The description to display.

$sel:enableDefaultInternetAccess:Fleet', fleet_enableDefaultInternetAccess - Indicates whether default internet access is enabled for the fleet.

$sel:fleetErrors:Fleet', fleet_fleetErrors - The fleet errors.

$sel:displayName:Fleet', fleet_displayName - The fleet name to display.

$sel:imageArn:Fleet', fleet_imageArn - The ARN for the public, private, or shared image.

$sel:arn:Fleet', fleet_arn - The Amazon Resource Name (ARN) for the fleet.

$sel:name:Fleet', fleet_name - The name of the fleet.

$sel:instanceType:Fleet', fleet_instanceType - The instance type to use when launching fleet instances. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

$sel:computeCapacityStatus:Fleet', fleet_computeCapacityStatus - The capacity status for the fleet.

$sel:state:Fleet', fleet_state - The current state for the fleet.

fleet_disconnectTimeoutInSeconds :: Lens' Fleet (Maybe Int)

The amount of time that a streaming session remains active after users disconnect. If they try to reconnect to the streaming session after a disconnection or network interruption within this time interval, they are connected to their previous session. Otherwise, they are connected to a new session with a new streaming instance.

Specify a value between 60 and 360000.

fleet_maxUserDurationInSeconds :: Lens' Fleet (Maybe Int)

The maximum amount of time that a streaming session can remain active, in seconds. If users are still connected to a streaming instance five minutes before this limit is reached, they are prompted to save any open documents before being disconnected. After this time elapses, the instance is terminated and replaced by a new instance.

Specify a value between 600 and 360000.

fleet_vpcConfig :: Lens' Fleet (Maybe VpcConfig)

The VPC configuration for the fleet.

fleet_iamRoleArn :: Lens' Fleet (Maybe Text)

The ARN of the IAM role that is applied to the fleet. To assume a role, the fleet instance calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

fleet_domainJoinInfo :: Lens' Fleet (Maybe DomainJoinInfo)

The name of the directory and organizational unit (OU) to use to join the fleet to a Microsoft Active Directory domain.

fleet_fleetType :: Lens' Fleet (Maybe FleetType)

The fleet type.

ALWAYS_ON

Provides users with instant-on access to their apps. You are charged for all running instances in your fleet, even if no users are streaming apps.

ON_DEMAND

Provide users with access to applications after they connect, which takes one to two minutes. You are charged for instance streaming when users are connected and a small hourly fee for instances that are not streaming apps.

fleet_idleDisconnectTimeoutInSeconds :: Lens' Fleet (Maybe Int)

The amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the DisconnectTimeoutInSeconds time interval begins. Users are notified before they are disconnected due to inactivity. If users try to reconnect to the streaming session before the time interval specified in DisconnectTimeoutInSeconds elapses, they are connected to their previous session. Users are considered idle when they stop providing keyboard or mouse input during their streaming session. File uploads and downloads, audio in, audio out, and pixels changing do not qualify as user activity. If users continue to be idle after the time interval in IdleDisconnectTimeoutInSeconds elapses, they are disconnected.

To prevent users from being disconnected due to inactivity, specify a value of 0. Otherwise, specify a value between 60 and 3600. The default value is 0.

If you enable this feature, we recommend that you specify a value that corresponds exactly to a whole number of minutes (for example, 60, 120, and 180). If you don't do this, the value is rounded to the nearest minute. For example, if you specify a value of 70, users are disconnected after 1 minute of inactivity. If you specify a value that is at the midpoint between two different minutes, the value is rounded up. For example, if you specify a value of 90, users are disconnected after 2 minutes of inactivity.

fleet_imageName :: Lens' Fleet (Maybe Text)

The name of the image used to create the fleet.

fleet_createdTime :: Lens' Fleet (Maybe UTCTime)

The time the fleet was created.

fleet_streamView :: Lens' Fleet (Maybe StreamView)

The AppStream 2.0 view that is displayed to your users when they stream from the fleet. When APP is specified, only the windows of applications opened by users display. When DESKTOP is specified, the standard desktop that is provided by the operating system displays.

The default value is APP.

fleet_description :: Lens' Fleet (Maybe Text)

The description to display.

fleet_enableDefaultInternetAccess :: Lens' Fleet (Maybe Bool)

Indicates whether default internet access is enabled for the fleet.

fleet_fleetErrors :: Lens' Fleet (Maybe [FleetError])

The fleet errors.

fleet_displayName :: Lens' Fleet (Maybe Text)

The fleet name to display.

fleet_imageArn :: Lens' Fleet (Maybe Text)

The ARN for the public, private, or shared image.

fleet_arn :: Lens' Fleet Text

The Amazon Resource Name (ARN) for the fleet.

fleet_name :: Lens' Fleet Text

The name of the fleet.

fleet_instanceType :: Lens' Fleet Text

The instance type to use when launching fleet instances. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

fleet_computeCapacityStatus :: Lens' Fleet ComputeCapacityStatus

The capacity status for the fleet.

fleet_state :: Lens' Fleet FleetState

The current state for the fleet.