Network.AWS.AppStream.Types.ImageBuilder

Description

Documentation

data ImageBuilder

Describes a virtual machine that is used to create an image.

See: newImageBuilder smart constructor.

Constructors

ImageBuilder'

Fields

vpcConfig :: Maybe VpcConfig

The VPC configuration of the image builder.

platform :: Maybe PlatformType

The operating system platform of the image builder.

iamRoleArn :: Maybe Text

The ARN of the IAM role that is applied to the image builder. To assume a role, the image builder calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

accessEndpoints :: Maybe (NonEmpty AccessEndpoint)

The list of virtual private cloud (VPC) interface endpoint objects. Administrators can connect to the image builder only through the specified endpoints.

domainJoinInfo :: Maybe DomainJoinInfo

The name of the directory and organizational unit (OU) to use to join the image builder to a Microsoft Active Directory domain.

instanceType :: Maybe Text

The instance type for the image builder. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

arn :: Maybe Text

The ARN for the image builder.

createdTime :: Maybe POSIX

The time stamp when the image builder was created.

stateChangeReason :: Maybe ImageBuilderStateChangeReason

The reason why the last state change occurred.

networkAccessConfiguration :: Maybe NetworkAccessConfiguration

state :: Maybe ImageBuilderState

The state of the image builder.

appstreamAgentVersion :: Maybe Text

The version of the AppStream 2.0 agent that is currently being used by the image builder.

description :: Maybe Text

The description to display.

enableDefaultInternetAccess :: Maybe Bool

Enables or disables default internet access for the image builder.

imageBuilderErrors :: Maybe [ResourceError]

The image builder errors.

displayName :: Maybe Text

The image builder name to display.

imageArn :: Maybe Text

The ARN of the image from which this builder was created.

name :: Text

The name of the image builder.

newImageBuilder

Arguments

:: Text
$sel:name:ImageBuilder'

-> ImageBuilder

Create a value of ImageBuilder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcConfig:ImageBuilder', imageBuilder_vpcConfig - The VPC configuration of the image builder.

$sel:platform:ImageBuilder', imageBuilder_platform - The operating system platform of the image builder.

$sel:iamRoleArn:ImageBuilder', imageBuilder_iamRoleArn - The ARN of the IAM role that is applied to the image builder. To assume a role, the image builder calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

$sel:accessEndpoints:ImageBuilder', imageBuilder_accessEndpoints - The list of virtual private cloud (VPC) interface endpoint objects. Administrators can connect to the image builder only through the specified endpoints.

$sel:domainJoinInfo:ImageBuilder', imageBuilder_domainJoinInfo - The name of the directory and organizational unit (OU) to use to join the image builder to a Microsoft Active Directory domain.

$sel:instanceType:ImageBuilder', imageBuilder_instanceType - The instance type for the image builder. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

$sel:arn:ImageBuilder', imageBuilder_arn - The ARN for the image builder.

$sel:createdTime:ImageBuilder', imageBuilder_createdTime - The time stamp when the image builder was created.

$sel:stateChangeReason:ImageBuilder', imageBuilder_stateChangeReason - The reason why the last state change occurred.

$sel:networkAccessConfiguration:ImageBuilder', imageBuilder_networkAccessConfiguration - Undocumented member.

$sel:state:ImageBuilder', imageBuilder_state - The state of the image builder.

$sel:appstreamAgentVersion:ImageBuilder', imageBuilder_appstreamAgentVersion - The version of the AppStream 2.0 agent that is currently being used by the image builder.

$sel:description:ImageBuilder', imageBuilder_description - The description to display.

$sel:enableDefaultInternetAccess:ImageBuilder', imageBuilder_enableDefaultInternetAccess - Enables or disables default internet access for the image builder.

$sel:imageBuilderErrors:ImageBuilder', imageBuilder_imageBuilderErrors - The image builder errors.

$sel:displayName:ImageBuilder', imageBuilder_displayName - The image builder name to display.

$sel:imageArn:ImageBuilder', imageBuilder_imageArn - The ARN of the image from which this builder was created.

$sel:name:ImageBuilder', imageBuilder_name - The name of the image builder.

imageBuilder_vpcConfig :: Lens' ImageBuilder (Maybe VpcConfig)

The VPC configuration of the image builder.

imageBuilder_platform :: Lens' ImageBuilder (Maybe PlatformType)

The operating system platform of the image builder.

imageBuilder_iamRoleArn :: Lens' ImageBuilder (Maybe Text)

The ARN of the IAM role that is applied to the image builder. To assume a role, the image builder calls the AWS Security Token Service (STS) AssumeRole API operation and passes the ARN of the role to use. The operation creates a new session with temporary credentials. AppStream 2.0 retrieves the temporary credentials and creates the appstream_machine_role credential profile on the instance.

For more information, see Using an IAM Role to Grant Permissions to Applications and Scripts Running on AppStream 2.0 Streaming Instances in the Amazon AppStream 2.0 Administration Guide.

imageBuilder_accessEndpoints :: Lens' ImageBuilder (Maybe (NonEmpty AccessEndpoint))

The list of virtual private cloud (VPC) interface endpoint objects. Administrators can connect to the image builder only through the specified endpoints.

imageBuilder_domainJoinInfo :: Lens' ImageBuilder (Maybe DomainJoinInfo)

The name of the directory and organizational unit (OU) to use to join the image builder to a Microsoft Active Directory domain.

imageBuilder_instanceType :: Lens' ImageBuilder (Maybe Text)

The instance type for the image builder. The following instance types are available:

stream.standard.small

stream.standard.medium

stream.standard.large

stream.compute.large

stream.compute.xlarge

stream.compute.2xlarge

stream.compute.4xlarge

stream.compute.8xlarge

stream.memory.large

stream.memory.xlarge

stream.memory.2xlarge

stream.memory.4xlarge

stream.memory.8xlarge

stream.memory.z1d.large

stream.memory.z1d.xlarge

stream.memory.z1d.2xlarge

stream.memory.z1d.3xlarge

stream.memory.z1d.6xlarge

stream.memory.z1d.12xlarge

stream.graphics-design.large

stream.graphics-design.xlarge

stream.graphics-design.2xlarge

stream.graphics-design.4xlarge

stream.graphics-desktop.2xlarge

stream.graphics.g4dn.xlarge

stream.graphics.g4dn.2xlarge

stream.graphics.g4dn.4xlarge

stream.graphics.g4dn.8xlarge

stream.graphics.g4dn.12xlarge

stream.graphics.g4dn.16xlarge

stream.graphics-pro.4xlarge

stream.graphics-pro.8xlarge

stream.graphics-pro.16xlarge

imageBuilder_arn :: Lens' ImageBuilder (Maybe Text)

The ARN for the image builder.

imageBuilder_createdTime :: Lens' ImageBuilder (Maybe UTCTime)

The time stamp when the image builder was created.

imageBuilder_stateChangeReason :: Lens' ImageBuilder (Maybe ImageBuilderStateChangeReason)

The reason why the last state change occurred.

imageBuilder_networkAccessConfiguration :: Lens' ImageBuilder (Maybe NetworkAccessConfiguration)

Undocumented member.

imageBuilder_state :: Lens' ImageBuilder (Maybe ImageBuilderState)

The state of the image builder.

imageBuilder_appstreamAgentVersion :: Lens' ImageBuilder (Maybe Text)

The version of the AppStream 2.0 agent that is currently being used by the image builder.

imageBuilder_description :: Lens' ImageBuilder (Maybe Text)

The description to display.

imageBuilder_enableDefaultInternetAccess :: Lens' ImageBuilder (Maybe Bool)

Enables or disables default internet access for the image builder.

imageBuilder_imageBuilderErrors :: Lens' ImageBuilder (Maybe [ResourceError])

The image builder errors.

imageBuilder_displayName :: Lens' ImageBuilder (Maybe Text)

The image builder name to display.

imageBuilder_imageArn :: Lens' ImageBuilder (Maybe Text)

The ARN of the image from which this builder was created.

imageBuilder_name :: Lens' ImageBuilder Text

The name of the image builder.