Network.AWS.AppStream.Types.ServiceAccountCredentials

Description

Documentation

data ServiceAccountCredentials

Describes the credentials for the service account used by the fleet or image builder to connect to the directory.

See: newServiceAccountCredentials smart constructor.

Constructors

ServiceAccountCredentials'

Fields

accountName :: Sensitive Text

The user name of the account. This account must have the following privileges: create computer objects, join computers to the domain, and change/reset the password on descendant computer objects for the organizational units specified.

accountPassword :: Sensitive Text

The password for the account.

newServiceAccountCredentials

Arguments

:: Text
$sel:accountName:ServiceAccountCredentials'

-> Text
$sel:accountPassword:ServiceAccountCredentials'

-> ServiceAccountCredentials

Create a value of ServiceAccountCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountName:ServiceAccountCredentials', serviceAccountCredentials_accountName - The user name of the account. This account must have the following privileges: create computer objects, join computers to the domain, and change/reset the password on descendant computer objects for the organizational units specified.

$sel:accountPassword:ServiceAccountCredentials', serviceAccountCredentials_accountPassword - The password for the account.

serviceAccountCredentials_accountName :: Lens' ServiceAccountCredentials Text

The user name of the account. This account must have the following privileges: create computer objects, join computers to the domain, and change/reset the password on descendant computer objects for the organizational units specified.

serviceAccountCredentials_accountPassword :: Lens' ServiceAccountCredentials Text

The password for the account.