Network.AWS.AppStream.Types.Session

Description

Documentation

data Session

Describes a streaming session.

See: newSession smart constructor.

Constructors

Session'

Fields

networkAccessConfiguration :: Maybe NetworkAccessConfiguration

The network details for the streaming session.

maxExpirationTime :: Maybe POSIX

The time when the streaming session is set to expire. This time is based on the MaxUserDurationinSeconds value, which determines the maximum length of time that a streaming session can run. A streaming session might end earlier than the time specified in SessionMaxExpirationTime, when the DisconnectTimeOutInSeconds elapses or the user chooses to end his or her session. If the DisconnectTimeOutInSeconds elapses, or the user chooses to end his or her session, the streaming instance is terminated and the streaming session ends.

startTime :: Maybe POSIX

The time when a streaming instance is dedicated for the user.

authenticationType :: Maybe AuthenticationType

The authentication method. The user is authenticated using a streaming URL (API) or SAML 2.0 federation (SAML).

connectionState :: Maybe SessionConnectionState

Specifies whether a user is connected to the streaming session.

id :: Text

The identifier of the streaming session.

userId :: Text

The identifier of the user for whom the session was created.

stackName :: Text

The name of the stack for the streaming session.

fleetName :: Text

The name of the fleet for the streaming session.

state :: SessionState

The current state of the streaming session.

newSession

Arguments

:: Text
$sel:id:Session'

-> Text
$sel:userId:Session'

-> Text
$sel:stackName:Session'

-> Text
$sel:fleetName:Session'

-> SessionState
$sel:state:Session'

-> Session

Create a value of Session with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkAccessConfiguration:Session', session_networkAccessConfiguration - The network details for the streaming session.

$sel:maxExpirationTime:Session', session_maxExpirationTime - The time when the streaming session is set to expire. This time is based on the MaxUserDurationinSeconds value, which determines the maximum length of time that a streaming session can run. A streaming session might end earlier than the time specified in SessionMaxExpirationTime, when the DisconnectTimeOutInSeconds elapses or the user chooses to end his or her session. If the DisconnectTimeOutInSeconds elapses, or the user chooses to end his or her session, the streaming instance is terminated and the streaming session ends.

$sel:startTime:Session', session_startTime - The time when a streaming instance is dedicated for the user.

$sel:authenticationType:Session', session_authenticationType - The authentication method. The user is authenticated using a streaming URL (API) or SAML 2.0 federation (SAML).

$sel:connectionState:Session', session_connectionState - Specifies whether a user is connected to the streaming session.

$sel:id:Session', session_id - The identifier of the streaming session.

$sel:userId:Session', session_userId - The identifier of the user for whom the session was created.

$sel:stackName:Session', session_stackName - The name of the stack for the streaming session.

$sel:fleetName:Session', session_fleetName - The name of the fleet for the streaming session.

$sel:state:Session', session_state - The current state of the streaming session.

session_networkAccessConfiguration :: Lens' Session (Maybe NetworkAccessConfiguration)

The network details for the streaming session.

session_maxExpirationTime :: Lens' Session (Maybe UTCTime)

The time when the streaming session is set to expire. This time is based on the MaxUserDurationinSeconds value, which determines the maximum length of time that a streaming session can run. A streaming session might end earlier than the time specified in SessionMaxExpirationTime, when the DisconnectTimeOutInSeconds elapses or the user chooses to end his or her session. If the DisconnectTimeOutInSeconds elapses, or the user chooses to end his or her session, the streaming instance is terminated and the streaming session ends.

session_startTime :: Lens' Session (Maybe UTCTime)

The time when a streaming instance is dedicated for the user.

session_authenticationType :: Lens' Session (Maybe AuthenticationType)

The authentication method. The user is authenticated using a streaming URL (API) or SAML 2.0 federation (SAML).

session_connectionState :: Lens' Session (Maybe SessionConnectionState)

Specifies whether a user is connected to the streaming session.

session_id :: Lens' Session Text

The identifier of the streaming session.

session_userId :: Lens' Session Text

The identifier of the user for whom the session was created.

session_stackName :: Lens' Session Text

The name of the stack for the streaming session.

session_fleetName :: Lens' Session Text

The name of the fleet for the streaming session.

session_state :: Lens' Session SessionState

The current state of the streaming session.