Network.AWS.AppStream.Types.User

Description

Documentation

data User

Describes a user in the user pool.

See: newUser smart constructor.

Constructors

User'

Fields

status :: Maybe Text

The status of the user in the user pool. The status can be one of the following:

UNCONFIRMED – The user is created but not confirmed.

CONFIRMED – The user is confirmed.

ARCHIVED – The user is no longer active.

COMPROMISED – The user is disabled because of a potential security threat.

UNKNOWN – The user status is not known.

arn :: Maybe Text

The ARN of the user.

enabled :: Maybe Bool

Specifies whether the user in the user pool is enabled.

createdTime :: Maybe POSIX

The date and time the user was created in the user pool.

userName :: Maybe (Sensitive Text)

The email address of the user.

Users' email addresses are case-sensitive.

lastName :: Maybe (Sensitive Text)

The last name, or surname, of the user.

firstName :: Maybe (Sensitive Text)

The first name, or given name, of the user.

authenticationType :: AuthenticationType

The authentication type for the user.

newUser

Arguments

:: AuthenticationType
$sel:authenticationType:User'

-> User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:User', user_status - The status of the user in the user pool. The status can be one of the following:

UNCONFIRMED – The user is created but not confirmed.

CONFIRMED – The user is confirmed.

ARCHIVED – The user is no longer active.

COMPROMISED – The user is disabled because of a potential security threat.

UNKNOWN – The user status is not known.

$sel:arn:User', user_arn - The ARN of the user.

$sel:enabled:User', user_enabled - Specifies whether the user in the user pool is enabled.

$sel:createdTime:User', user_createdTime - The date and time the user was created in the user pool.

$sel:userName:User', user_userName - The email address of the user.

Users' email addresses are case-sensitive.

$sel:lastName:User', user_lastName - The last name, or surname, of the user.

$sel:firstName:User', user_firstName - The first name, or given name, of the user.

$sel:authenticationType:User', user_authenticationType - The authentication type for the user.

user_status :: Lens' User (Maybe Text)

The status of the user in the user pool. The status can be one of the following:

UNCONFIRMED – The user is created but not confirmed.

CONFIRMED – The user is confirmed.

ARCHIVED – The user is no longer active.

COMPROMISED – The user is disabled because of a potential security threat.

UNKNOWN – The user status is not known.

user_arn :: Lens' User (Maybe Text)

The ARN of the user.

user_enabled :: Lens' User (Maybe Bool)

Specifies whether the user in the user pool is enabled.

user_createdTime :: Lens' User (Maybe UTCTime)

The date and time the user was created in the user pool.

user_userName :: Lens' User (Maybe Text)

The email address of the user.

Users' email addresses are case-sensitive.

user_lastName :: Lens' User (Maybe Text)

The last name, or surname, of the user.

user_firstName :: Lens' User (Maybe Text)

The first name, or given name, of the user.

user_authenticationType :: Lens' User AuthenticationType

The authentication type for the user.