Network.AWS.AppStream.Types.UserStackAssociation

Description

Documentation

data UserStackAssociation

Describes a user in the user pool and the associated stack.

See: newUserStackAssociation smart constructor.

Constructors

UserStackAssociation'

Fields

sendEmailNotification :: Maybe Bool

Specifies whether a welcome email is sent to a user after the user is created in the user pool.

stackName :: Text

The name of the stack that is associated with the user.

userName :: Sensitive Text

The email address of the user who is associated with the stack.

Users' email addresses are case-sensitive.

authenticationType :: AuthenticationType

The authentication type for the user.

newUserStackAssociation

Arguments

:: Text
$sel:stackName:UserStackAssociation'

-> Text
$sel:userName:UserStackAssociation'

-> AuthenticationType
$sel:authenticationType:UserStackAssociation'

-> UserStackAssociation

Create a value of UserStackAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sendEmailNotification:UserStackAssociation', userStackAssociation_sendEmailNotification - Specifies whether a welcome email is sent to a user after the user is created in the user pool.

$sel:stackName:UserStackAssociation', userStackAssociation_stackName - The name of the stack that is associated with the user.

$sel:userName:UserStackAssociation', userStackAssociation_userName - The email address of the user who is associated with the stack.

Users' email addresses are case-sensitive.

$sel:authenticationType:UserStackAssociation', userStackAssociation_authenticationType - The authentication type for the user.

userStackAssociation_sendEmailNotification :: Lens' UserStackAssociation (Maybe Bool)

Specifies whether a welcome email is sent to a user after the user is created in the user pool.

userStackAssociation_stackName :: Lens' UserStackAssociation Text

The name of the stack that is associated with the user.

userStackAssociation_userName :: Lens' UserStackAssociation Text

The email address of the user who is associated with the stack.

Users' email addresses are case-sensitive.

userStackAssociation_authenticationType :: Lens' UserStackAssociation AuthenticationType

The authentication type for the user.