Network.AWS.AppStream.UpdateDirectoryConfig

Description

Updates the specified Directory Config object in AppStream 2.0. This object includes the configuration information required to join fleets and image builders to Microsoft Active Directory domains.

Creating a Request

data UpdateDirectoryConfig

See: newUpdateDirectoryConfig smart constructor.

Constructors

UpdateDirectoryConfig'

Fields

serviceAccountCredentials :: Maybe ServiceAccountCredentials

The credentials for the service account used by the fleet or image builder to connect to the directory.

organizationalUnitDistinguishedNames :: Maybe [Text]

The distinguished names of the organizational units for computer accounts.

directoryName :: Text

The name of the Directory Config object.

newUpdateDirectoryConfig

Arguments

:: Text
$sel:directoryName:UpdateDirectoryConfig'

-> UpdateDirectoryConfig

Create a value of UpdateDirectoryConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceAccountCredentials:UpdateDirectoryConfig', updateDirectoryConfig_serviceAccountCredentials - The credentials for the service account used by the fleet or image builder to connect to the directory.

$sel:organizationalUnitDistinguishedNames:UpdateDirectoryConfig', updateDirectoryConfig_organizationalUnitDistinguishedNames - The distinguished names of the organizational units for computer accounts.

$sel:directoryName:UpdateDirectoryConfig', updateDirectoryConfig_directoryName - The name of the Directory Config object.

Request Lenses

updateDirectoryConfig_serviceAccountCredentials :: Lens' UpdateDirectoryConfig (Maybe ServiceAccountCredentials)

The credentials for the service account used by the fleet or image builder to connect to the directory.

updateDirectoryConfig_organizationalUnitDistinguishedNames :: Lens' UpdateDirectoryConfig (Maybe [Text])

The distinguished names of the organizational units for computer accounts.

updateDirectoryConfig_directoryName :: Lens' UpdateDirectoryConfig Text

The name of the Directory Config object.

Destructuring the Response

data UpdateDirectoryConfigResponse

See: newUpdateDirectoryConfigResponse smart constructor.

Constructors

UpdateDirectoryConfigResponse'

Fields

directoryConfig :: Maybe DirectoryConfig

Information about the Directory Config object.

httpStatus :: Int

The response's http status code.

newUpdateDirectoryConfigResponse

Arguments

:: Int
$sel:httpStatus:UpdateDirectoryConfigResponse'

-> UpdateDirectoryConfigResponse

Create a value of UpdateDirectoryConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryConfig:UpdateDirectoryConfigResponse', updateDirectoryConfigResponse_directoryConfig - Information about the Directory Config object.

$sel:httpStatus:UpdateDirectoryConfigResponse', updateDirectoryConfigResponse_httpStatus - The response's http status code.

Response Lenses

updateDirectoryConfigResponse_directoryConfig :: Lens' UpdateDirectoryConfigResponse (Maybe DirectoryConfig)

Information about the Directory Config object.

updateDirectoryConfigResponse_httpStatus :: Lens' UpdateDirectoryConfigResponse Int

The response's http status code.