Network.AWS.AppStream.UpdateStack

Description

Updates the specified fields for the specified stack.

Creating a Request

data UpdateStack

See: newUpdateStack smart constructor.

Constructors

UpdateStack'

Fields

userSettings :: Maybe (NonEmpty UserSetting)

The actions that are enabled or disabled for users during their streaming sessions. By default, these actions are enabled.

accessEndpoints :: Maybe (NonEmpty AccessEndpoint)

The list of interface VPC endpoint (interface endpoint) objects. Users of the stack can connect to AppStream 2.0 only through the specified endpoints.

redirectURL :: Maybe Text

The URL that users are redirected to after their streaming session ends.

applicationSettings :: Maybe ApplicationSettings

The persistent application settings for users of a stack. When these settings are enabled, changes that users make to applications and Windows settings are automatically saved after each session and applied to the next session.

storageConnectors :: Maybe [StorageConnector]

The storage connectors to enable.

description :: Maybe Text

The description to display.

deleteStorageConnectors :: Maybe Bool

Deletes the storage connectors currently enabled for the stack.

embedHostDomains :: Maybe (NonEmpty Text)

The domains where AppStream 2.0 streaming sessions can be embedded in an iframe. You must approve the domains that you want to host embedded AppStream 2.0 streaming sessions.

attributesToDelete :: Maybe [StackAttribute]

The stack attributes to delete.

displayName :: Maybe Text

The stack name to display.

feedbackURL :: Maybe Text

The URL that users are redirected to after they choose the Send Feedback link. If no URL is specified, no Send Feedback link is displayed.

name :: Text

The name of the stack.

newUpdateStack

Arguments

:: Text
$sel:name:UpdateStack'

-> UpdateStack

Create a value of UpdateStack with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userSettings:UpdateStack', updateStack_userSettings - The actions that are enabled or disabled for users during their streaming sessions. By default, these actions are enabled.

$sel:accessEndpoints:UpdateStack', updateStack_accessEndpoints - The list of interface VPC endpoint (interface endpoint) objects. Users of the stack can connect to AppStream 2.0 only through the specified endpoints.

$sel:redirectURL:UpdateStack', updateStack_redirectURL - The URL that users are redirected to after their streaming session ends.

$sel:applicationSettings:UpdateStack', updateStack_applicationSettings - The persistent application settings for users of a stack. When these settings are enabled, changes that users make to applications and Windows settings are automatically saved after each session and applied to the next session.

$sel:storageConnectors:UpdateStack', updateStack_storageConnectors - The storage connectors to enable.

$sel:description:UpdateStack', updateStack_description - The description to display.

$sel:deleteStorageConnectors:UpdateStack', updateStack_deleteStorageConnectors - Deletes the storage connectors currently enabled for the stack.

$sel:embedHostDomains:UpdateStack', updateStack_embedHostDomains - The domains where AppStream 2.0 streaming sessions can be embedded in an iframe. You must approve the domains that you want to host embedded AppStream 2.0 streaming sessions.

$sel:attributesToDelete:UpdateStack', updateStack_attributesToDelete - The stack attributes to delete.

$sel:displayName:UpdateStack', updateStack_displayName - The stack name to display.

$sel:feedbackURL:UpdateStack', updateStack_feedbackURL - The URL that users are redirected to after they choose the Send Feedback link. If no URL is specified, no Send Feedback link is displayed.

$sel:name:UpdateStack', updateStack_name - The name of the stack.

Request Lenses

updateStack_userSettings :: Lens' UpdateStack (Maybe (NonEmpty UserSetting))

The actions that are enabled or disabled for users during their streaming sessions. By default, these actions are enabled.

updateStack_accessEndpoints :: Lens' UpdateStack (Maybe (NonEmpty AccessEndpoint))

The list of interface VPC endpoint (interface endpoint) objects. Users of the stack can connect to AppStream 2.0 only through the specified endpoints.

updateStack_redirectURL :: Lens' UpdateStack (Maybe Text)

The URL that users are redirected to after their streaming session ends.

updateStack_applicationSettings :: Lens' UpdateStack (Maybe ApplicationSettings)

The persistent application settings for users of a stack. When these settings are enabled, changes that users make to applications and Windows settings are automatically saved after each session and applied to the next session.

updateStack_storageConnectors :: Lens' UpdateStack (Maybe [StorageConnector])

The storage connectors to enable.

updateStack_description :: Lens' UpdateStack (Maybe Text)

The description to display.

updateStack_deleteStorageConnectors :: Lens' UpdateStack (Maybe Bool)

Deletes the storage connectors currently enabled for the stack.

updateStack_embedHostDomains :: Lens' UpdateStack (Maybe (NonEmpty Text))

The domains where AppStream 2.0 streaming sessions can be embedded in an iframe. You must approve the domains that you want to host embedded AppStream 2.0 streaming sessions.

updateStack_attributesToDelete :: Lens' UpdateStack (Maybe [StackAttribute])

The stack attributes to delete.

updateStack_displayName :: Lens' UpdateStack (Maybe Text)

The stack name to display.

updateStack_feedbackURL :: Lens' UpdateStack (Maybe Text)

The URL that users are redirected to after they choose the Send Feedback link. If no URL is specified, no Send Feedback link is displayed.

updateStack_name :: Lens' UpdateStack Text

The name of the stack.

Destructuring the Response

data UpdateStackResponse

See: newUpdateStackResponse smart constructor.

Constructors

UpdateStackResponse'

Fields

stack :: Maybe Stack

Information about the stack.

httpStatus :: Int

The response's http status code.

newUpdateStackResponse

Arguments

:: Int
$sel:httpStatus:UpdateStackResponse'

-> UpdateStackResponse

Create a value of UpdateStackResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stack:UpdateStackResponse', updateStackResponse_stack - Information about the stack.

$sel:httpStatus:UpdateStackResponse', updateStackResponse_httpStatus - The response's http status code.

Response Lenses

updateStackResponse_stack :: Lens' UpdateStackResponse (Maybe Stack)

Information about the stack.

updateStackResponse_httpStatus :: Lens' UpdateStackResponse Int

The response's http status code.