Network.AWS.AppSync.CreateDataSource

Description

Creates a DataSource object.

Creating a Request

data CreateDataSource

See: newCreateDataSource smart constructor.

Constructors

CreateDataSource'

Fields

serviceRoleArn :: Maybe Text

The Identity and Access Management service role ARN for the data source. The system assumes this role when accessing the data source.

relationalDatabaseConfig :: Maybe RelationalDatabaseDataSourceConfig

Relational database settings.

elasticsearchConfig :: Maybe ElasticsearchDataSourceConfig

Amazon OpenSearch Service settings.

As of September 2021, Amazon Elasticsearch service is Amazon OpenSearch Service. This configuration is deprecated. For new data sources, use CreateDataSourceRequest$openSearchServiceConfig to create an OpenSearch data source.

lambdaConfig :: Maybe LambdaDataSourceConfig

Amazon Web Services Lambda settings.

dynamodbConfig :: Maybe DynamodbDataSourceConfig

Amazon DynamoDB settings.

description :: Maybe Text

A description of the DataSource.

openSearchServiceConfig :: Maybe OpenSearchServiceDataSourceConfig

Amazon OpenSearch Service settings.

httpConfig :: Maybe HttpDataSourceConfig

HTTP endpoint settings.

apiId :: Text

The API ID for the GraphQL API for the DataSource.

name :: Text

A user-supplied name for the DataSource.

type' :: DataSourceType

The type of the DataSource.

newCreateDataSource

Arguments

:: Text
$sel:apiId:CreateDataSource'

-> Text
$sel:name:CreateDataSource'

-> DataSourceType
$sel:type':CreateDataSource'

-> CreateDataSource

Create a value of CreateDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceRoleArn:CreateDataSource', createDataSource_serviceRoleArn - The Identity and Access Management service role ARN for the data source. The system assumes this role when accessing the data source.

$sel:relationalDatabaseConfig:CreateDataSource', createDataSource_relationalDatabaseConfig - Relational database settings.

$sel:elasticsearchConfig:CreateDataSource', createDataSource_elasticsearchConfig - Amazon OpenSearch Service settings.

As of September 2021, Amazon Elasticsearch service is Amazon OpenSearch Service. This configuration is deprecated. For new data sources, use CreateDataSourceRequest$openSearchServiceConfig to create an OpenSearch data source.

$sel:lambdaConfig:CreateDataSource', createDataSource_lambdaConfig - Amazon Web Services Lambda settings.

$sel:dynamodbConfig:CreateDataSource', createDataSource_dynamodbConfig - Amazon DynamoDB settings.

$sel:description:CreateDataSource', createDataSource_description - A description of the DataSource.

$sel:openSearchServiceConfig:CreateDataSource', createDataSource_openSearchServiceConfig - Amazon OpenSearch Service settings.

$sel:httpConfig:CreateDataSource', createDataSource_httpConfig - HTTP endpoint settings.

$sel:apiId:CreateDataSource', createDataSource_apiId - The API ID for the GraphQL API for the DataSource.

$sel:name:CreateDataSource', createDataSource_name - A user-supplied name for the DataSource.

$sel:type':CreateDataSource', createDataSource_type - The type of the DataSource.

Request Lenses

createDataSource_serviceRoleArn :: Lens' CreateDataSource (Maybe Text)

The Identity and Access Management service role ARN for the data source. The system assumes this role when accessing the data source.

createDataSource_relationalDatabaseConfig :: Lens' CreateDataSource (Maybe RelationalDatabaseDataSourceConfig)

Relational database settings.

createDataSource_elasticsearchConfig :: Lens' CreateDataSource (Maybe ElasticsearchDataSourceConfig)

Amazon OpenSearch Service settings.

As of September 2021, Amazon Elasticsearch service is Amazon OpenSearch Service. This configuration is deprecated. For new data sources, use CreateDataSourceRequest$openSearchServiceConfig to create an OpenSearch data source.

createDataSource_lambdaConfig :: Lens' CreateDataSource (Maybe LambdaDataSourceConfig)

Amazon Web Services Lambda settings.

createDataSource_dynamodbConfig :: Lens' CreateDataSource (Maybe DynamodbDataSourceConfig)

Amazon DynamoDB settings.

createDataSource_description :: Lens' CreateDataSource (Maybe Text)

A description of the DataSource.

createDataSource_openSearchServiceConfig :: Lens' CreateDataSource (Maybe OpenSearchServiceDataSourceConfig)

Amazon OpenSearch Service settings.

createDataSource_httpConfig :: Lens' CreateDataSource (Maybe HttpDataSourceConfig)

HTTP endpoint settings.

createDataSource_apiId :: Lens' CreateDataSource Text

The API ID for the GraphQL API for the DataSource.

createDataSource_name :: Lens' CreateDataSource Text

A user-supplied name for the DataSource.

createDataSource_type :: Lens' CreateDataSource DataSourceType

The type of the DataSource.

Destructuring the Response

data CreateDataSourceResponse

See: newCreateDataSourceResponse smart constructor.

Constructors

CreateDataSourceResponse'

Fields

dataSource :: Maybe DataSource

The DataSource object.

httpStatus :: Int

The response's http status code.

newCreateDataSourceResponse

Arguments

:: Int
$sel:httpStatus:CreateDataSourceResponse'

-> CreateDataSourceResponse

Create a value of CreateDataSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSource:CreateDataSourceResponse', createDataSourceResponse_dataSource - The DataSource object.

$sel:httpStatus:CreateDataSourceResponse', createDataSourceResponse_httpStatus - The response's http status code.

Response Lenses

createDataSourceResponse_dataSource :: Lens' CreateDataSourceResponse (Maybe DataSource)

The DataSource object.

createDataSourceResponse_httpStatus :: Lens' CreateDataSourceResponse Int

The response's http status code.