Network.AWS.AppSync.CreateGraphqlApi

Description

Creates a GraphqlApi object.

Creating a Request

data CreateGraphqlApi

See: newCreateGraphqlApi smart constructor.

Constructors

CreateGraphqlApi'

Fields

openIDConnectConfig :: Maybe OpenIDConnectConfig

The OpenID Connect configuration.

userPoolConfig :: Maybe UserPoolConfig

The Amazon Cognito user pool configuration.

xrayEnabled :: Maybe Bool

A flag indicating whether to enable X-Ray tracing for the GraphqlApi.

lambdaAuthorizerConfig :: Maybe LambdaAuthorizerConfig

Configuration for Amazon Web Services Lambda function authorization.

tags :: Maybe (HashMap Text Text)

A TagMap object.

logConfig :: Maybe LogConfig

The Amazon CloudWatch Logs configuration.

additionalAuthenticationProviders :: Maybe [AdditionalAuthenticationProvider]

A list of additional authentication providers for the GraphqlApi API.

name :: Text

A user-supplied name for the GraphqlApi.

authenticationType :: AuthenticationType

The authentication type: API key, Identity and Access Management, OIDC, Amazon Cognito user pools, or Amazon Web Services Lambda.

newCreateGraphqlApi

Arguments

:: Text
$sel:name:CreateGraphqlApi'

-> AuthenticationType
$sel:authenticationType:CreateGraphqlApi'

-> CreateGraphqlApi

Create a value of CreateGraphqlApi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:openIDConnectConfig:CreateGraphqlApi', createGraphqlApi_openIDConnectConfig - The OpenID Connect configuration.

$sel:userPoolConfig:CreateGraphqlApi', createGraphqlApi_userPoolConfig - The Amazon Cognito user pool configuration.

$sel:xrayEnabled:CreateGraphqlApi', createGraphqlApi_xrayEnabled - A flag indicating whether to enable X-Ray tracing for the GraphqlApi.

$sel:lambdaAuthorizerConfig:CreateGraphqlApi', createGraphqlApi_lambdaAuthorizerConfig - Configuration for Amazon Web Services Lambda function authorization.

$sel:tags:CreateGraphqlApi', createGraphqlApi_tags - A TagMap object.

$sel:logConfig:CreateGraphqlApi', createGraphqlApi_logConfig - The Amazon CloudWatch Logs configuration.

$sel:additionalAuthenticationProviders:CreateGraphqlApi', createGraphqlApi_additionalAuthenticationProviders - A list of additional authentication providers for the GraphqlApi API.

$sel:name:CreateGraphqlApi', createGraphqlApi_name - A user-supplied name for the GraphqlApi.

$sel:authenticationType:CreateGraphqlApi', createGraphqlApi_authenticationType - The authentication type: API key, Identity and Access Management, OIDC, Amazon Cognito user pools, or Amazon Web Services Lambda.

Request Lenses

createGraphqlApi_openIDConnectConfig :: Lens' CreateGraphqlApi (Maybe OpenIDConnectConfig)

The OpenID Connect configuration.

createGraphqlApi_userPoolConfig :: Lens' CreateGraphqlApi (Maybe UserPoolConfig)

The Amazon Cognito user pool configuration.

createGraphqlApi_xrayEnabled :: Lens' CreateGraphqlApi (Maybe Bool)

A flag indicating whether to enable X-Ray tracing for the GraphqlApi.

createGraphqlApi_lambdaAuthorizerConfig :: Lens' CreateGraphqlApi (Maybe LambdaAuthorizerConfig)

Configuration for Amazon Web Services Lambda function authorization.

createGraphqlApi_tags :: Lens' CreateGraphqlApi (Maybe (HashMap Text Text))

A TagMap object.

createGraphqlApi_logConfig :: Lens' CreateGraphqlApi (Maybe LogConfig)

The Amazon CloudWatch Logs configuration.

createGraphqlApi_additionalAuthenticationProviders :: Lens' CreateGraphqlApi (Maybe [AdditionalAuthenticationProvider])

A list of additional authentication providers for the GraphqlApi API.

createGraphqlApi_name :: Lens' CreateGraphqlApi Text

A user-supplied name for the GraphqlApi.

createGraphqlApi_authenticationType :: Lens' CreateGraphqlApi AuthenticationType

The authentication type: API key, Identity and Access Management, OIDC, Amazon Cognito user pools, or Amazon Web Services Lambda.

Destructuring the Response

data CreateGraphqlApiResponse

See: newCreateGraphqlApiResponse smart constructor.

Constructors

CreateGraphqlApiResponse'

Fields

graphqlApi :: Maybe GraphqlApi

The GraphqlApi.

httpStatus :: Int

The response's http status code.

newCreateGraphqlApiResponse

Arguments

:: Int
$sel:httpStatus:CreateGraphqlApiResponse'

-> CreateGraphqlApiResponse

Create a value of CreateGraphqlApiResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:graphqlApi:CreateGraphqlApiResponse', createGraphqlApiResponse_graphqlApi - The GraphqlApi.

$sel:httpStatus:CreateGraphqlApiResponse', createGraphqlApiResponse_httpStatus - The response's http status code.

Response Lenses

createGraphqlApiResponse_graphqlApi :: Lens' CreateGraphqlApiResponse (Maybe GraphqlApi)

The GraphqlApi.

createGraphqlApiResponse_httpStatus :: Lens' CreateGraphqlApiResponse Int

The response's http status code.