Network.AWS.AppSync.CreateResolver

Description

Creates a Resolver object.

A resolver converts incoming requests into a format that a data source can understand and converts the data source's responses into GraphQL.

Creating a Request

data CreateResolver

See: newCreateResolver smart constructor.

Constructors

CreateResolver'

Fields

responseMappingTemplate :: Maybe Text

The mapping template to be used for responses from the data source.

kind :: Maybe ResolverKind

The resolver type.

UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source.

PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

syncConfig :: Maybe SyncConfig

The SyncConfig for a resolver attached to a versioned datasource.

dataSourceName :: Maybe Text

The name of the data source for which the resolver is being created.

cachingConfig :: Maybe CachingConfig

The caching configuration for the resolver.

pipelineConfig :: Maybe PipelineConfig

The PipelineConfig.

requestMappingTemplate :: Maybe Text

The mapping template to be used for requests.

A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL).

VTL request mapping templates are optional when using a Lambda data source. For all other data sources, VTL request and response mapping templates are required.

apiId :: Text

The ID for the GraphQL API for which the resolver is being created.

typeName :: Text

The name of the Type.

fieldName :: Text

The name of the field to attach the resolver to.

newCreateResolver

Arguments

:: Text
$sel:apiId:CreateResolver'

-> Text
$sel:typeName:CreateResolver'

-> Text
$sel:fieldName:CreateResolver'

-> CreateResolver

Create a value of CreateResolver with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:responseMappingTemplate:CreateResolver', createResolver_responseMappingTemplate - The mapping template to be used for responses from the data source.

$sel:kind:CreateResolver', createResolver_kind - The resolver type.

UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source.

PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

$sel:syncConfig:CreateResolver', createResolver_syncConfig - The SyncConfig for a resolver attached to a versioned datasource.

$sel:dataSourceName:CreateResolver', createResolver_dataSourceName - The name of the data source for which the resolver is being created.

$sel:cachingConfig:CreateResolver', createResolver_cachingConfig - The caching configuration for the resolver.

$sel:pipelineConfig:CreateResolver', createResolver_pipelineConfig - The PipelineConfig.

$sel:requestMappingTemplate:CreateResolver', createResolver_requestMappingTemplate - The mapping template to be used for requests.

A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL).

VTL request mapping templates are optional when using a Lambda data source. For all other data sources, VTL request and response mapping templates are required.

$sel:apiId:CreateResolver', createResolver_apiId - The ID for the GraphQL API for which the resolver is being created.

$sel:typeName:CreateResolver', createResolver_typeName - The name of the Type.

$sel:fieldName:CreateResolver', createResolver_fieldName - The name of the field to attach the resolver to.

Request Lenses

createResolver_responseMappingTemplate :: Lens' CreateResolver (Maybe Text)

The mapping template to be used for responses from the data source.

createResolver_kind :: Lens' CreateResolver (Maybe ResolverKind)

The resolver type.

UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source.

PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

createResolver_syncConfig :: Lens' CreateResolver (Maybe SyncConfig)

The SyncConfig for a resolver attached to a versioned datasource.

createResolver_dataSourceName :: Lens' CreateResolver (Maybe Text)

The name of the data source for which the resolver is being created.

createResolver_cachingConfig :: Lens' CreateResolver (Maybe CachingConfig)

The caching configuration for the resolver.

createResolver_pipelineConfig :: Lens' CreateResolver (Maybe PipelineConfig)

The PipelineConfig.

createResolver_requestMappingTemplate :: Lens' CreateResolver (Maybe Text)

The mapping template to be used for requests.

A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL).

VTL request mapping templates are optional when using a Lambda data source. For all other data sources, VTL request and response mapping templates are required.

createResolver_apiId :: Lens' CreateResolver Text

The ID for the GraphQL API for which the resolver is being created.

createResolver_typeName :: Lens' CreateResolver Text

The name of the Type.

createResolver_fieldName :: Lens' CreateResolver Text

The name of the field to attach the resolver to.

Destructuring the Response

data CreateResolverResponse

See: newCreateResolverResponse smart constructor.

Constructors

CreateResolverResponse'

Fields

resolver :: Maybe Resolver

The Resolver object.

httpStatus :: Int

The response's http status code.

newCreateResolverResponse

Arguments

:: Int
$sel:httpStatus:CreateResolverResponse'

-> CreateResolverResponse

Create a value of CreateResolverResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolver:CreateResolverResponse', createResolverResponse_resolver - The Resolver object.

$sel:httpStatus:CreateResolverResponse', createResolverResponse_httpStatus - The response's http status code.

Response Lenses

createResolverResponse_resolver :: Lens' CreateResolverResponse (Maybe Resolver)

The Resolver object.

createResolverResponse_httpStatus :: Lens' CreateResolverResponse Int

The response's http status code.