Network.AWS.AppSync.CreateType

Description

Creates a Type object.

Creating a Request

data CreateType

See: newCreateType smart constructor.

Constructors

CreateType'

Fields

apiId :: Text

The API ID.

definition :: Text

The type definition, in GraphQL Schema Definition Language (SDL) format.

For more information, see the GraphQL SDL documentation.

format :: TypeDefinitionFormat

The type format: SDL or JSON.

newCreateType

Arguments

:: Text
$sel:apiId:CreateType'

-> Text
$sel:definition:CreateType'

-> TypeDefinitionFormat
$sel:format:CreateType'

-> CreateType

Create a value of CreateType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:CreateType', createType_apiId - The API ID.

$sel:definition:CreateType', createType_definition - The type definition, in GraphQL Schema Definition Language (SDL) format.

For more information, see the GraphQL SDL documentation.

$sel:format:CreateType', createType_format - The type format: SDL or JSON.

Request Lenses

createType_apiId :: Lens' CreateType Text

The API ID.

createType_definition :: Lens' CreateType Text

The type definition, in GraphQL Schema Definition Language (SDL) format.

For more information, see the GraphQL SDL documentation.

createType_format :: Lens' CreateType TypeDefinitionFormat

The type format: SDL or JSON.

Destructuring the Response

data CreateTypeResponse

See: newCreateTypeResponse smart constructor.

Constructors

CreateTypeResponse'

Fields

type' :: Maybe Type

The Type object.

httpStatus :: Int

The response's http status code.

newCreateTypeResponse

Arguments

:: Int
$sel:httpStatus:CreateTypeResponse'

-> CreateTypeResponse

Create a value of CreateTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':CreateTypeResponse', createTypeResponse_type - The Type object.

$sel:httpStatus:CreateTypeResponse', createTypeResponse_httpStatus - The response's http status code.

Response Lenses

createTypeResponse_type :: Lens' CreateTypeResponse (Maybe Type)

The Type object.

createTypeResponse_httpStatus :: Lens' CreateTypeResponse Int

The response's http status code.