Network.AWS.AppSync.DeleteApiCache

Description

Deletes an ApiCache object.

Creating a Request

data DeleteApiCache

Represents the input of a DeleteApiCache operation.

See: newDeleteApiCache smart constructor.

Constructors

DeleteApiCache'

Fields

apiId :: Text

The API ID.

newDeleteApiCache

Arguments

:: Text
$sel:apiId:DeleteApiCache'

-> DeleteApiCache

Create a value of DeleteApiCache with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:DeleteApiCache', deleteApiCache_apiId - The API ID.

Request Lenses

deleteApiCache_apiId :: Lens' DeleteApiCache Text

The API ID.

Destructuring the Response

data DeleteApiCacheResponse

Represents the output of a DeleteApiCache operation.

See: newDeleteApiCacheResponse smart constructor.

Constructors

DeleteApiCacheResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteApiCacheResponse

Arguments

:: Int
$sel:httpStatus:DeleteApiCacheResponse'

-> DeleteApiCacheResponse

Create a value of DeleteApiCacheResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteApiCacheResponse', deleteApiCacheResponse_httpStatus - The response's http status code.

Response Lenses

deleteApiCacheResponse_httpStatus :: Lens' DeleteApiCacheResponse Int

The response's http status code.