Network.AWS.AppSync.GetDataSource

Description

Retrieves a DataSource object.

Creating a Request

data GetDataSource

See: newGetDataSource smart constructor.

Constructors

GetDataSource'

Fields

apiId :: Text

The API ID.

name :: Text

The name of the data source.

newGetDataSource

Arguments

:: Text
$sel:apiId:GetDataSource'

-> Text
$sel:name:GetDataSource'

-> GetDataSource

Create a value of GetDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:GetDataSource', getDataSource_apiId - The API ID.

$sel:name:GetDataSource', getDataSource_name - The name of the data source.

Request Lenses

getDataSource_apiId :: Lens' GetDataSource Text

The API ID.

getDataSource_name :: Lens' GetDataSource Text

The name of the data source.

Destructuring the Response

data GetDataSourceResponse

See: newGetDataSourceResponse smart constructor.

Constructors

GetDataSourceResponse'

Fields

dataSource :: Maybe DataSource

The DataSource object.

httpStatus :: Int

The response's http status code.

newGetDataSourceResponse

Arguments

:: Int
$sel:httpStatus:GetDataSourceResponse'

-> GetDataSourceResponse

Create a value of GetDataSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSource:GetDataSourceResponse', getDataSourceResponse_dataSource - The DataSource object.

$sel:httpStatus:GetDataSourceResponse', getDataSourceResponse_httpStatus - The response's http status code.

Response Lenses

getDataSourceResponse_dataSource :: Lens' GetDataSourceResponse (Maybe DataSource)

The DataSource object.

getDataSourceResponse_httpStatus :: Lens' GetDataSourceResponse Int

The response's http status code.