Network.AWS.AppSync.Lens

Description

Operations

ListResolvers

listResolvers_nextToken :: Lens' ListResolvers (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listResolvers_maxResults :: Lens' ListResolvers (Maybe Natural)

The maximum number of results you want the request to return.

listResolvers_apiId :: Lens' ListResolvers Text

The API ID.

listResolvers_typeName :: Lens' ListResolvers Text

The type name.

listResolversResponse_nextToken :: Lens' ListResolversResponse (Maybe Text)

An identifier to be passed in the next request to this operation to return the next set of items in the list.

listResolversResponse_resolvers :: Lens' ListResolversResponse (Maybe [Resolver])

The Resolver objects.

listResolversResponse_httpStatus :: Lens' ListResolversResponse Int

The response's http status code.

UpdateDataSource

updateDataSource_serviceRoleArn :: Lens' UpdateDataSource (Maybe Text)

The new service role ARN for the data source.

updateDataSource_relationalDatabaseConfig :: Lens' UpdateDataSource (Maybe RelationalDatabaseDataSourceConfig)

The new relational database configuration.

updateDataSource_elasticsearchConfig :: Lens' UpdateDataSource (Maybe ElasticsearchDataSourceConfig)

The new OpenSearch configuration.

As of September 2021, Amazon Elasticsearch service is Amazon OpenSearch Service. This configuration is deprecated. Instead, use UpdateDataSourceRequest$openSearchServiceConfig to update an OpenSearch data source.

updateDataSource_lambdaConfig :: Lens' UpdateDataSource (Maybe LambdaDataSourceConfig)

The new Amazon Web Services Lambda configuration.

updateDataSource_dynamodbConfig :: Lens' UpdateDataSource (Maybe DynamodbDataSourceConfig)

The new Amazon DynamoDB configuration.

updateDataSource_description :: Lens' UpdateDataSource (Maybe Text)

The new description for the data source.

updateDataSource_openSearchServiceConfig :: Lens' UpdateDataSource (Maybe OpenSearchServiceDataSourceConfig)

The new OpenSearch configuration.

updateDataSource_httpConfig :: Lens' UpdateDataSource (Maybe HttpDataSourceConfig)

The new HTTP endpoint configuration.

updateDataSource_apiId :: Lens' UpdateDataSource Text

The API ID.

updateDataSource_name :: Lens' UpdateDataSource Text

The new name for the data source.

updateDataSource_type :: Lens' UpdateDataSource DataSourceType

The new data source type.

updateDataSourceResponse_dataSource :: Lens' UpdateDataSourceResponse (Maybe DataSource)

The updated DataSource object.

updateDataSourceResponse_httpStatus :: Lens' UpdateDataSourceResponse Int

The response's http status code.

ListDataSources

listDataSources_nextToken :: Lens' ListDataSources (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listDataSources_maxResults :: Lens' ListDataSources (Maybe Natural)

The maximum number of results you want the request to return.

listDataSources_apiId :: Lens' ListDataSources Text

The API ID.

listDataSourcesResponse_nextToken :: Lens' ListDataSourcesResponse (Maybe Text)

An identifier to be passed in the next request to this operation to return the next set of items in the list.

listDataSourcesResponse_dataSources :: Lens' ListDataSourcesResponse (Maybe [DataSource])

The DataSource objects.

listDataSourcesResponse_httpStatus :: Lens' ListDataSourcesResponse Int

The response's http status code.

GetFunction

getFunction_apiId :: Lens' GetFunction Text

The GraphQL API ID.

getFunction_functionId :: Lens' GetFunction Text

The Function ID.

getFunctionResponse_functionConfiguration :: Lens' GetFunctionResponse (Maybe FunctionConfiguration)

The Function object.

getFunctionResponse_httpStatus :: Lens' GetFunctionResponse Int

The response's http status code.

CreateType

createType_apiId :: Lens' CreateType Text

The API ID.

createType_definition :: Lens' CreateType Text

The type definition, in GraphQL Schema Definition Language (SDL) format.

For more information, see the GraphQL SDL documentation.

createType_format :: Lens' CreateType TypeDefinitionFormat

The type format: SDL or JSON.

createTypeResponse_type :: Lens' CreateTypeResponse (Maybe Type)

The Type object.

createTypeResponse_httpStatus :: Lens' CreateTypeResponse Int

The response's http status code.

DeleteDataSource

deleteDataSource_apiId :: Lens' DeleteDataSource Text

The API ID.

deleteDataSource_name :: Lens' DeleteDataSource Text

The name of the data source.

deleteDataSourceResponse_httpStatus :: Lens' DeleteDataSourceResponse Int

The response's http status code.

UpdateType

updateType_definition :: Lens' UpdateType (Maybe Text)

The new definition.

updateType_apiId :: Lens' UpdateType Text

The API ID.

updateType_typeName :: Lens' UpdateType Text

The new type name.

updateType_format :: Lens' UpdateType TypeDefinitionFormat

The new type format: SDL or JSON.

updateTypeResponse_type :: Lens' UpdateTypeResponse (Maybe Type)

The updated Type object.

updateTypeResponse_httpStatus :: Lens' UpdateTypeResponse Int

The response's http status code.

CreateDataSource

createDataSource_serviceRoleArn :: Lens' CreateDataSource (Maybe Text)

The Identity and Access Management service role ARN for the data source. The system assumes this role when accessing the data source.

createDataSource_relationalDatabaseConfig :: Lens' CreateDataSource (Maybe RelationalDatabaseDataSourceConfig)

Relational database settings.

createDataSource_elasticsearchConfig :: Lens' CreateDataSource (Maybe ElasticsearchDataSourceConfig)

Amazon OpenSearch Service settings.

As of September 2021, Amazon Elasticsearch service is Amazon OpenSearch Service. This configuration is deprecated. For new data sources, use CreateDataSourceRequest$openSearchServiceConfig to create an OpenSearch data source.

createDataSource_lambdaConfig :: Lens' CreateDataSource (Maybe LambdaDataSourceConfig)

Amazon Web Services Lambda settings.

createDataSource_dynamodbConfig :: Lens' CreateDataSource (Maybe DynamodbDataSourceConfig)

Amazon DynamoDB settings.

createDataSource_description :: Lens' CreateDataSource (Maybe Text)

A description of the DataSource.

createDataSource_openSearchServiceConfig :: Lens' CreateDataSource (Maybe OpenSearchServiceDataSourceConfig)

Amazon OpenSearch Service settings.

createDataSource_httpConfig :: Lens' CreateDataSource (Maybe HttpDataSourceConfig)

HTTP endpoint settings.

createDataSource_apiId :: Lens' CreateDataSource Text

The API ID for the GraphQL API for the DataSource.

createDataSource_name :: Lens' CreateDataSource Text

A user-supplied name for the DataSource.

createDataSource_type :: Lens' CreateDataSource DataSourceType

The type of the DataSource.

createDataSourceResponse_dataSource :: Lens' CreateDataSourceResponse (Maybe DataSource)

The DataSource object.

createDataSourceResponse_httpStatus :: Lens' CreateDataSourceResponse Int

The response's http status code.

DeleteType

deleteType_apiId :: Lens' DeleteType Text

The API ID.

deleteType_typeName :: Lens' DeleteType Text

The type name.

deleteTypeResponse_httpStatus :: Lens' DeleteTypeResponse Int

The response's http status code.

ListResolversByFunction

listResolversByFunction_nextToken :: Lens' ListResolversByFunction (Maybe Text)

An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.

listResolversByFunction_maxResults :: Lens' ListResolversByFunction (Maybe Natural)

The maximum number of results you want the request to return.

listResolversByFunction_apiId :: Lens' ListResolversByFunction Text

The API ID.

listResolversByFunction_functionId :: Lens' ListResolversByFunction Text

The Function ID.

listResolversByFunctionResponse_nextToken :: Lens' ListResolversByFunctionResponse (Maybe Text)

An identifier that can be used to return the next set of items in the list.

listResolversByFunctionResponse_resolvers :: Lens' ListResolversByFunctionResponse (Maybe [Resolver])

The list of resolvers.

listResolversByFunctionResponse_httpStatus :: Lens' ListResolversByFunctionResponse Int

The response's http status code.

ListFunctions

listFunctions_nextToken :: Lens' ListFunctions (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listFunctions_maxResults :: Lens' ListFunctions (Maybe Natural)

The maximum number of results you want the request to return.

listFunctions_apiId :: Lens' ListFunctions Text

The GraphQL API ID.

listFunctionsResponse_nextToken :: Lens' ListFunctionsResponse (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listFunctionsResponse_functions :: Lens' ListFunctionsResponse (Maybe [FunctionConfiguration])

A list of Function objects.

listFunctionsResponse_httpStatus :: Lens' ListFunctionsResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The GraphqlApi ARN.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

A list of TagKey objects.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DeleteGraphqlApi

deleteGraphqlApi_apiId :: Lens' DeleteGraphqlApi Text

The API ID.

deleteGraphqlApiResponse_httpStatus :: Lens' DeleteGraphqlApiResponse Int

The response's http status code.

GetResolver

getResolver_apiId :: Lens' GetResolver Text

The API ID.

getResolver_typeName :: Lens' GetResolver Text

The resolver type name.

getResolver_fieldName :: Lens' GetResolver Text

The resolver field name.

getResolverResponse_resolver :: Lens' GetResolverResponse (Maybe Resolver)

The Resolver object.

getResolverResponse_httpStatus :: Lens' GetResolverResponse Int

The response's http status code.

GetIntrospectionSchema

getIntrospectionSchema_includeDirectives :: Lens' GetIntrospectionSchema (Maybe Bool)

A flag that specifies whether the schema introspection should contain directives.

getIntrospectionSchema_apiId :: Lens' GetIntrospectionSchema Text

The API ID.

getIntrospectionSchema_format :: Lens' GetIntrospectionSchema OutputType

The schema format: SDL or JSON.

getIntrospectionSchemaResponse_schema :: Lens' GetIntrospectionSchemaResponse (Maybe ByteString)

The schema, in GraphQL Schema Definition Language (SDL) format.

For more information, see the GraphQL SDL documentation.

getIntrospectionSchemaResponse_httpStatus :: Lens' GetIntrospectionSchemaResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The GraphqlApi ARN.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

A TagMap object.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListGraphqlApis

listGraphqlApis_nextToken :: Lens' ListGraphqlApis (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listGraphqlApis_maxResults :: Lens' ListGraphqlApis (Maybe Natural)

The maximum number of results you want the request to return.

listGraphqlApisResponse_nextToken :: Lens' ListGraphqlApisResponse (Maybe Text)

An identifier to be passed in the next request to this operation to return the next set of items in the list.

listGraphqlApisResponse_graphqlApis :: Lens' ListGraphqlApisResponse (Maybe [GraphqlApi])

The GraphqlApi objects.

listGraphqlApisResponse_httpStatus :: Lens' ListGraphqlApisResponse Int

The response's http status code.

UpdateGraphqlApi

updateGraphqlApi_openIDConnectConfig :: Lens' UpdateGraphqlApi (Maybe OpenIDConnectConfig)

The OpenID Connect configuration for the GraphqlApi object.

updateGraphqlApi_userPoolConfig :: Lens' UpdateGraphqlApi (Maybe UserPoolConfig)

The new Amazon Cognito user pool configuration for the GraphqlApi object.

updateGraphqlApi_xrayEnabled :: Lens' UpdateGraphqlApi (Maybe Bool)

A flag indicating whether to enable X-Ray tracing for the GraphqlApi.

updateGraphqlApi_lambdaAuthorizerConfig :: Lens' UpdateGraphqlApi (Maybe LambdaAuthorizerConfig)

Configuration for Amazon Web Services Lambda function authorization.

updateGraphqlApi_logConfig :: Lens' UpdateGraphqlApi (Maybe LogConfig)

The Amazon CloudWatch Logs configuration for the GraphqlApi object.

updateGraphqlApi_additionalAuthenticationProviders :: Lens' UpdateGraphqlApi (Maybe [AdditionalAuthenticationProvider])

A list of additional authentication providers for the GraphqlApi API.

updateGraphqlApi_authenticationType :: Lens' UpdateGraphqlApi (Maybe AuthenticationType)

The new authentication type for the GraphqlApi object.

updateGraphqlApi_apiId :: Lens' UpdateGraphqlApi Text

The API ID.

updateGraphqlApi_name :: Lens' UpdateGraphqlApi Text

The new name for the GraphqlApi object.

updateGraphqlApiResponse_graphqlApi :: Lens' UpdateGraphqlApiResponse (Maybe GraphqlApi)

The updated GraphqlApi object.

updateGraphqlApiResponse_httpStatus :: Lens' UpdateGraphqlApiResponse Int

The response's http status code.

UpdateApiCache

updateApiCache_apiId :: Lens' UpdateApiCache Text

The GraphQL API Id.

updateApiCache_ttl :: Lens' UpdateApiCache Integer

TTL in seconds for cache entries.

Valid values are between 1 and 3600 seconds.

updateApiCache_apiCachingBehavior :: Lens' UpdateApiCache ApiCachingBehavior

Caching behavior.

FULL_REQUEST_CACHING: All requests are fully cached.

PER_RESOLVER_CACHING: Individual resolvers that you specify are cached.

updateApiCache_type :: Lens' UpdateApiCache ApiCacheType

The cache instance type. Valid values are

SMALL

MEDIUM

LARGE

XLARGE

LARGE_2X

LARGE_4X

LARGE_8X (not available in all regions)

LARGE_12X

Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used.

The following legacy instance types are available, but their use is discouraged:

T2_SMALL: A t2.small instance type.

T2_MEDIUM: A t2.medium instance type.

R4_LARGE: A r4.large instance type.

R4_XLARGE: A r4.xlarge instance type.

R4_2XLARGE: A r4.2xlarge instance type.

R4_4XLARGE: A r4.4xlarge instance type.

R4_8XLARGE: A r4.8xlarge instance type.

updateApiCacheResponse_apiCache :: Lens' UpdateApiCacheResponse (Maybe ApiCache)

The ApiCache object.

updateApiCacheResponse_httpStatus :: Lens' UpdateApiCacheResponse Int

The response's http status code.

DeleteApiCache

deleteApiCache_apiId :: Lens' DeleteApiCache Text

The API ID.

deleteApiCacheResponse_httpStatus :: Lens' DeleteApiCacheResponse Int

The response's http status code.

GetSchemaCreationStatus

getSchemaCreationStatus_apiId :: Lens' GetSchemaCreationStatus Text

The API ID.

getSchemaCreationStatusResponse_status :: Lens' GetSchemaCreationStatusResponse (Maybe SchemaStatus)

The current state of the schema (PROCESSING, FAILED, SUCCESS, or NOT_APPLICABLE). When the schema is in the ACTIVE state, you can add data.

getSchemaCreationStatusResponse_details :: Lens' GetSchemaCreationStatusResponse (Maybe Text)

Detailed information about the status of the schema creation operation.

getSchemaCreationStatusResponse_httpStatus :: Lens' GetSchemaCreationStatusResponse Int

The response's http status code.

GetApiCache

getApiCache_apiId :: Lens' GetApiCache Text

The API ID.

getApiCacheResponse_apiCache :: Lens' GetApiCacheResponse (Maybe ApiCache)

The ApiCache object.

getApiCacheResponse_httpStatus :: Lens' GetApiCacheResponse Int

The response's http status code.

GetGraphqlApi

getGraphqlApi_apiId :: Lens' GetGraphqlApi Text

The API ID for the GraphQL API.

getGraphqlApiResponse_graphqlApi :: Lens' GetGraphqlApiResponse (Maybe GraphqlApi)

The GraphqlApi object.

getGraphqlApiResponse_httpStatus :: Lens' GetGraphqlApiResponse Int

The response's http status code.

UpdateResolver

updateResolver_responseMappingTemplate :: Lens' UpdateResolver (Maybe Text)

The new response mapping template.

updateResolver_kind :: Lens' UpdateResolver (Maybe ResolverKind)

The resolver type.

UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source.

PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

updateResolver_syncConfig :: Lens' UpdateResolver (Maybe SyncConfig)

The SyncConfig for a resolver attached to a versioned datasource.

updateResolver_dataSourceName :: Lens' UpdateResolver (Maybe Text)

The new data source name.

updateResolver_cachingConfig :: Lens' UpdateResolver (Maybe CachingConfig)

The caching configuration for the resolver.

updateResolver_pipelineConfig :: Lens' UpdateResolver (Maybe PipelineConfig)

The PipelineConfig.

updateResolver_requestMappingTemplate :: Lens' UpdateResolver (Maybe Text)

The new request mapping template.

A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL).

VTL request mapping templates are optional when using a Lambda data source. For all other data sources, VTL request and response mapping templates are required.

updateResolver_apiId :: Lens' UpdateResolver Text

The API ID.

updateResolver_typeName :: Lens' UpdateResolver Text

The new type name.

updateResolver_fieldName :: Lens' UpdateResolver Text

The new field name.

updateResolverResponse_resolver :: Lens' UpdateResolverResponse (Maybe Resolver)

The updated Resolver object.

updateResolverResponse_httpStatus :: Lens' UpdateResolverResponse Int

The response's http status code.

DeleteResolver

deleteResolver_apiId :: Lens' DeleteResolver Text

The API ID.

deleteResolver_typeName :: Lens' DeleteResolver Text

The name of the resolver type.

deleteResolver_fieldName :: Lens' DeleteResolver Text

The resolver field name.

deleteResolverResponse_httpStatus :: Lens' DeleteResolverResponse Int

The response's http status code.

ListTypes

listTypes_nextToken :: Lens' ListTypes (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listTypes_maxResults :: Lens' ListTypes (Maybe Natural)

The maximum number of results you want the request to return.

listTypes_apiId :: Lens' ListTypes Text

The API ID.

listTypes_format :: Lens' ListTypes TypeDefinitionFormat

The type format: SDL or JSON.

listTypesResponse_nextToken :: Lens' ListTypesResponse (Maybe Text)

An identifier to be passed in the next request to this operation to return the next set of items in the list.

listTypesResponse_types :: Lens' ListTypesResponse (Maybe [Type])

The Type objects.

listTypesResponse_httpStatus :: Lens' ListTypesResponse Int

The response's http status code.

GetType

getType_apiId :: Lens' GetType Text

The API ID.

getType_typeName :: Lens' GetType Text

The type name.

getType_format :: Lens' GetType TypeDefinitionFormat

The type format: SDL or JSON.

getTypeResponse_type :: Lens' GetTypeResponse (Maybe Type)

The Type object.

getTypeResponse_httpStatus :: Lens' GetTypeResponse Int

The response's http status code.

CreateFunction

createFunction_responseMappingTemplate :: Lens' CreateFunction (Maybe Text)

The Function response mapping template.

createFunction_syncConfig :: Lens' CreateFunction (Maybe SyncConfig)

Undocumented member.

createFunction_description :: Lens' CreateFunction (Maybe Text)

The Function description.

createFunction_requestMappingTemplate :: Lens' CreateFunction (Maybe Text)

The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

createFunction_apiId :: Lens' CreateFunction Text

The GraphQL API ID.

createFunction_name :: Lens' CreateFunction Text

The Function name. The function name does not have to be unique.

createFunction_dataSourceName :: Lens' CreateFunction Text

The Function DataSource name.

createFunction_functionVersion :: Lens' CreateFunction Text

The version of the request mapping template. Currently the supported value is 2018-05-29.

createFunctionResponse_functionConfiguration :: Lens' CreateFunctionResponse (Maybe FunctionConfiguration)

The Function object.

createFunctionResponse_httpStatus :: Lens' CreateFunctionResponse Int

The response's http status code.

DeleteApiKey

deleteApiKey_apiId :: Lens' DeleteApiKey Text

The API ID.

deleteApiKey_id :: Lens' DeleteApiKey Text

The ID for the API key.

deleteApiKeyResponse_httpStatus :: Lens' DeleteApiKeyResponse Int

The response's http status code.

ListApiKeys

listApiKeys_nextToken :: Lens' ListApiKeys (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listApiKeys_maxResults :: Lens' ListApiKeys (Maybe Natural)

The maximum number of results you want the request to return.

listApiKeys_apiId :: Lens' ListApiKeys Text

The API ID.

listApiKeysResponse_nextToken :: Lens' ListApiKeysResponse (Maybe Text)

An identifier to be passed in the next request to this operation to return the next set of items in the list.

listApiKeysResponse_apiKeys :: Lens' ListApiKeysResponse (Maybe [ApiKey])

The ApiKey objects.

listApiKeysResponse_httpStatus :: Lens' ListApiKeysResponse Int

The response's http status code.

UpdateApiKey

updateApiKey_description :: Lens' UpdateApiKey (Maybe Text)

A description of the purpose of the API key.

updateApiKey_expires :: Lens' UpdateApiKey (Maybe Integer)

The time from update time after which the API key expires. The date is represented as seconds since the epoch. For more information, see .

updateApiKey_apiId :: Lens' UpdateApiKey Text

The ID for the GraphQL API.

updateApiKey_id :: Lens' UpdateApiKey Text

The API key ID.

updateApiKeyResponse_apiKey :: Lens' UpdateApiKeyResponse (Maybe ApiKey)

The API key.

updateApiKeyResponse_httpStatus :: Lens' UpdateApiKeyResponse Int

The response's http status code.

CreateApiKey

createApiKey_description :: Lens' CreateApiKey (Maybe Text)

A description of the purpose of the API key.

createApiKey_expires :: Lens' CreateApiKey (Maybe Integer)

The time from creation time after which the API key expires. The date is represented as seconds since the epoch, rounded down to the nearest hour. The default value for this parameter is 7 days from creation time. For more information, see .

createApiKey_apiId :: Lens' CreateApiKey Text

The ID for your GraphQL API.

createApiKeyResponse_apiKey :: Lens' CreateApiKeyResponse (Maybe ApiKey)

The API key.

createApiKeyResponse_httpStatus :: Lens' CreateApiKeyResponse Int

The response's http status code.

DeleteFunction

deleteFunction_apiId :: Lens' DeleteFunction Text

The GraphQL API ID.

deleteFunction_functionId :: Lens' DeleteFunction Text

The Function ID.

deleteFunctionResponse_httpStatus :: Lens' DeleteFunctionResponse Int

The response's http status code.

UpdateFunction

updateFunction_responseMappingTemplate :: Lens' UpdateFunction (Maybe Text)

The Function request mapping template.

updateFunction_syncConfig :: Lens' UpdateFunction (Maybe SyncConfig)

Undocumented member.

updateFunction_description :: Lens' UpdateFunction (Maybe Text)

The Function description.

updateFunction_requestMappingTemplate :: Lens' UpdateFunction (Maybe Text)

The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

updateFunction_apiId :: Lens' UpdateFunction Text

The GraphQL API ID.

updateFunction_name :: Lens' UpdateFunction Text

The Function name.

updateFunction_functionId :: Lens' UpdateFunction Text

The function ID.

updateFunction_dataSourceName :: Lens' UpdateFunction Text

The Function DataSource name.

updateFunction_functionVersion :: Lens' UpdateFunction Text

The version of the request mapping template. Currently the supported value is 2018-05-29.

updateFunctionResponse_functionConfiguration :: Lens' UpdateFunctionResponse (Maybe FunctionConfiguration)

The Function object.

updateFunctionResponse_httpStatus :: Lens' UpdateFunctionResponse Int

The response's http status code.

GetDataSource

getDataSource_apiId :: Lens' GetDataSource Text

The API ID.

getDataSource_name :: Lens' GetDataSource Text

The name of the data source.

getDataSourceResponse_dataSource :: Lens' GetDataSourceResponse (Maybe DataSource)

The DataSource object.

getDataSourceResponse_httpStatus :: Lens' GetDataSourceResponse Int

The response's http status code.

StartSchemaCreation

startSchemaCreation_apiId :: Lens' StartSchemaCreation Text

The API ID.

startSchemaCreation_definition :: Lens' StartSchemaCreation ByteString

The schema definition, in GraphQL schema language format.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

startSchemaCreationResponse_status :: Lens' StartSchemaCreationResponse (Maybe SchemaStatus)

The current state of the schema (PROCESSING, FAILED, SUCCESS, or NOT_APPLICABLE). When the schema is in the ACTIVE state, you can add data.

startSchemaCreationResponse_httpStatus :: Lens' StartSchemaCreationResponse Int

The response's http status code.

CreateApiCache

createApiCache_atRestEncryptionEnabled :: Lens' CreateApiCache (Maybe Bool)

At rest encryption flag for cache. This setting cannot be updated after creation.

createApiCache_transitEncryptionEnabled :: Lens' CreateApiCache (Maybe Bool)

Transit encryption flag when connecting to cache. This setting cannot be updated after creation.

createApiCache_apiId :: Lens' CreateApiCache Text

The GraphQL API Id.

createApiCache_ttl :: Lens' CreateApiCache Integer

TTL in seconds for cache entries.

Valid values are between 1 and 3600 seconds.

createApiCache_apiCachingBehavior :: Lens' CreateApiCache ApiCachingBehavior

Caching behavior.

FULL_REQUEST_CACHING: All requests are fully cached.

PER_RESOLVER_CACHING: Individual resolvers that you specify are cached.

createApiCache_type :: Lens' CreateApiCache ApiCacheType

The cache instance type. Valid values are

SMALL

MEDIUM

LARGE

XLARGE

LARGE_2X

LARGE_4X

LARGE_8X (not available in all regions)

LARGE_12X

Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used.

The following legacy instance types are available, but their use is discouraged:

T2_SMALL: A t2.small instance type.

T2_MEDIUM: A t2.medium instance type.

R4_LARGE: A r4.large instance type.

R4_XLARGE: A r4.xlarge instance type.

R4_2XLARGE: A r4.2xlarge instance type.

R4_4XLARGE: A r4.4xlarge instance type.

R4_8XLARGE: A r4.8xlarge instance type.

createApiCacheResponse_apiCache :: Lens' CreateApiCacheResponse (Maybe ApiCache)

The ApiCache object.

createApiCacheResponse_httpStatus :: Lens' CreateApiCacheResponse Int

The response's http status code.

FlushApiCache

flushApiCache_apiId :: Lens' FlushApiCache Text

The API ID.

flushApiCacheResponse_httpStatus :: Lens' FlushApiCacheResponse Int

The response's http status code.

CreateGraphqlApi

createGraphqlApi_openIDConnectConfig :: Lens' CreateGraphqlApi (Maybe OpenIDConnectConfig)

The OpenID Connect configuration.

createGraphqlApi_userPoolConfig :: Lens' CreateGraphqlApi (Maybe UserPoolConfig)

The Amazon Cognito user pool configuration.

createGraphqlApi_xrayEnabled :: Lens' CreateGraphqlApi (Maybe Bool)

A flag indicating whether to enable X-Ray tracing for the GraphqlApi.

createGraphqlApi_lambdaAuthorizerConfig :: Lens' CreateGraphqlApi (Maybe LambdaAuthorizerConfig)

Configuration for Amazon Web Services Lambda function authorization.

createGraphqlApi_tags :: Lens' CreateGraphqlApi (Maybe (HashMap Text Text))

A TagMap object.

createGraphqlApi_logConfig :: Lens' CreateGraphqlApi (Maybe LogConfig)

The Amazon CloudWatch Logs configuration.

createGraphqlApi_additionalAuthenticationProviders :: Lens' CreateGraphqlApi (Maybe [AdditionalAuthenticationProvider])

A list of additional authentication providers for the GraphqlApi API.

createGraphqlApi_name :: Lens' CreateGraphqlApi Text

A user-supplied name for the GraphqlApi.

createGraphqlApi_authenticationType :: Lens' CreateGraphqlApi AuthenticationType

The authentication type: API key, Identity and Access Management, OIDC, Amazon Cognito user pools, or Amazon Web Services Lambda.

createGraphqlApiResponse_graphqlApi :: Lens' CreateGraphqlApiResponse (Maybe GraphqlApi)

The GraphqlApi.

createGraphqlApiResponse_httpStatus :: Lens' CreateGraphqlApiResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The GraphqlApi ARN.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

A TagMap object.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CreateResolver

createResolver_responseMappingTemplate :: Lens' CreateResolver (Maybe Text)

The mapping template to be used for responses from the data source.

createResolver_kind :: Lens' CreateResolver (Maybe ResolverKind)

The resolver type.

UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source.

PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

createResolver_syncConfig :: Lens' CreateResolver (Maybe SyncConfig)

The SyncConfig for a resolver attached to a versioned datasource.

createResolver_dataSourceName :: Lens' CreateResolver (Maybe Text)

The name of the data source for which the resolver is being created.

createResolver_cachingConfig :: Lens' CreateResolver (Maybe CachingConfig)

The caching configuration for the resolver.

createResolver_pipelineConfig :: Lens' CreateResolver (Maybe PipelineConfig)

The PipelineConfig.

createResolver_requestMappingTemplate :: Lens' CreateResolver (Maybe Text)

The mapping template to be used for requests.

A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL).

VTL request mapping templates are optional when using a Lambda data source. For all other data sources, VTL request and response mapping templates are required.

createResolver_apiId :: Lens' CreateResolver Text

The ID for the GraphQL API for which the resolver is being created.

createResolver_typeName :: Lens' CreateResolver Text

The name of the Type.

createResolver_fieldName :: Lens' CreateResolver Text

The name of the field to attach the resolver to.

createResolverResponse_resolver :: Lens' CreateResolverResponse (Maybe Resolver)

The Resolver object.

createResolverResponse_httpStatus :: Lens' CreateResolverResponse Int

The response's http status code.

Types

AdditionalAuthenticationProvider

additionalAuthenticationProvider_openIDConnectConfig :: Lens' AdditionalAuthenticationProvider (Maybe OpenIDConnectConfig)

The OpenID Connect configuration.

additionalAuthenticationProvider_userPoolConfig :: Lens' AdditionalAuthenticationProvider (Maybe CognitoUserPoolConfig)

The Amazon Cognito user pool configuration.

additionalAuthenticationProvider_lambdaAuthorizerConfig :: Lens' AdditionalAuthenticationProvider (Maybe LambdaAuthorizerConfig)

Configuration for Amazon Web Services Lambda function authorization.

additionalAuthenticationProvider_authenticationType :: Lens' AdditionalAuthenticationProvider (Maybe AuthenticationType)

The authentication type: API key, Identity and Access Management, OIDC, Amazon Cognito user pools, or Amazon Web Services Lambda.

ApiCache

apiCache_status :: Lens' ApiCache (Maybe ApiCacheStatus)

The cache instance status.

AVAILABLE: The instance is available for use.

CREATING: The instance is currently creating.

DELETING: The instance is currently deleting.

MODIFYING: The instance is currently modifying.

FAILED: The instance has failed creation.

apiCache_atRestEncryptionEnabled :: Lens' ApiCache (Maybe Bool)

At rest encryption flag for cache. This setting cannot be updated after creation.

apiCache_ttl :: Lens' ApiCache (Maybe Integer)

TTL in seconds for cache entries.

Valid values are between 1 and 3600 seconds.

apiCache_type :: Lens' ApiCache (Maybe ApiCacheType)

The cache instance type. Valid values are

SMALL

MEDIUM

LARGE

XLARGE

LARGE_2X

LARGE_4X

LARGE_8X (not available in all regions)

LARGE_12X

Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used.

The following legacy instance types are available, but their use is discouraged:

T2_SMALL: A t2.small instance type.

T2_MEDIUM: A t2.medium instance type.

R4_LARGE: A r4.large instance type.

R4_XLARGE: A r4.xlarge instance type.

R4_2XLARGE: A r4.2xlarge instance type.

R4_4XLARGE: A r4.4xlarge instance type.

R4_8XLARGE: A r4.8xlarge instance type.

apiCache_transitEncryptionEnabled :: Lens' ApiCache (Maybe Bool)

Transit encryption flag when connecting to cache. This setting cannot be updated after creation.

apiCache_apiCachingBehavior :: Lens' ApiCache (Maybe ApiCachingBehavior)

Caching behavior.

FULL_REQUEST_CACHING: All requests are fully cached.

PER_RESOLVER_CACHING: Individual resolvers that you specify are cached.

ApiKey

apiKey_id :: Lens' ApiKey (Maybe Text)

The API key ID.

apiKey_deletes :: Lens' ApiKey (Maybe Integer)

The time after which the API key is deleted. The date is represented as seconds since the epoch, rounded down to the nearest hour.

apiKey_description :: Lens' ApiKey (Maybe Text)

A description of the purpose of the API key.

apiKey_expires :: Lens' ApiKey (Maybe Integer)

The time after which the API key expires. The date is represented as seconds since the epoch, rounded down to the nearest hour.

AuthorizationConfig

authorizationConfig_awsIamConfig :: Lens' AuthorizationConfig (Maybe AwsIamConfig)

The Identity and Access Management settings.

authorizationConfig_authorizationType :: Lens' AuthorizationConfig AuthorizationType

The authorization type required by the HTTP endpoint.

AWS_IAM: The authorization type is Sigv4.

AwsIamConfig

awsIamConfig_signingServiceName :: Lens' AwsIamConfig (Maybe Text)

The signing service name for Identity and Access Management authorization.

awsIamConfig_signingRegion :: Lens' AwsIamConfig (Maybe Text)

The signing region for Identity and Access Management authorization.

CachingConfig

cachingConfig_ttl :: Lens' CachingConfig (Maybe Integer)

The TTL in seconds for a resolver that has caching enabled.

Valid values are between 1 and 3600 seconds.

cachingConfig_cachingKeys :: Lens' CachingConfig (Maybe [Text])

The caching keys for a resolver that has caching enabled.

Valid values are entries from the $context.arguments, $context.source, and $context.identity maps.

CognitoUserPoolConfig

cognitoUserPoolConfig_appIdClientRegex :: Lens' CognitoUserPoolConfig (Maybe Text)

A regular expression for validating the incoming Amazon Cognito user pool app client ID.

cognitoUserPoolConfig_userPoolId :: Lens' CognitoUserPoolConfig Text

The user pool ID.

cognitoUserPoolConfig_awsRegion :: Lens' CognitoUserPoolConfig Text

The Amazon Web Services Region in which the user pool was created.

DataSource

dataSource_serviceRoleArn :: Lens' DataSource (Maybe Text)

The Identity and Access Management service role ARN for the data source. The system assumes this role when accessing the data source.

dataSource_relationalDatabaseConfig :: Lens' DataSource (Maybe RelationalDatabaseDataSourceConfig)

Relational database settings.

dataSource_elasticsearchConfig :: Lens' DataSource (Maybe ElasticsearchDataSourceConfig)

Amazon OpenSearch Service settings.

dataSource_lambdaConfig :: Lens' DataSource (Maybe LambdaDataSourceConfig)

Amazon Web Services Lambda settings.

dataSource_name :: Lens' DataSource (Maybe Text)

The name of the data source.

dataSource_dynamodbConfig :: Lens' DataSource (Maybe DynamodbDataSourceConfig)

Amazon DynamoDB settings.

dataSource_description :: Lens' DataSource (Maybe Text)

The description of the data source.

dataSource_dataSourceArn :: Lens' DataSource (Maybe Text)

The data source ARN.

dataSource_type :: Lens' DataSource (Maybe DataSourceType)

The type of the data source.

AWS_LAMBDA: The data source is an Amazon Web Services Lambda function.

AMAZON_DYNAMODB: The data source is an Amazon DynamoDB table.

AMAZON_ELASTICSEARCH: The data source is an Amazon OpenSearch Service domain.

AMAZON_OPENSEARCH_SERVICE: The data source is an Amazon OpenSearch Service domain.

NONE: There is no data source. This type is used when you wish to invoke a GraphQL operation without connecting to a data source, such as performing data transformation with resolvers or triggering a subscription to be invoked from a mutation.

HTTP: The data source is an HTTP endpoint.

RELATIONAL_DATABASE: The data source is a relational database.

dataSource_openSearchServiceConfig :: Lens' DataSource (Maybe OpenSearchServiceDataSourceConfig)

Amazon OpenSearch Service settings.

dataSource_httpConfig :: Lens' DataSource (Maybe HttpDataSourceConfig)

HTTP endpoint settings.

DeltaSyncConfig

deltaSyncConfig_baseTableTTL :: Lens' DeltaSyncConfig (Maybe Integer)

The number of minutes an Item is stored in the datasource.

deltaSyncConfig_deltaSyncTableName :: Lens' DeltaSyncConfig (Maybe Text)

The Delta Sync table name.

deltaSyncConfig_deltaSyncTableTTL :: Lens' DeltaSyncConfig (Maybe Integer)

The number of minutes a Delta Sync log entry is stored in the Delta Sync table.

DynamodbDataSourceConfig

dynamodbDataSourceConfig_useCallerCredentials :: Lens' DynamodbDataSourceConfig (Maybe Bool)

Set to TRUE to use Amazon Cognito credentials with this data source.

dynamodbDataSourceConfig_versioned :: Lens' DynamodbDataSourceConfig (Maybe Bool)

Set to TRUE to use Conflict Detection and Resolution with this data source.

dynamodbDataSourceConfig_deltaSyncConfig :: Lens' DynamodbDataSourceConfig (Maybe DeltaSyncConfig)

The DeltaSyncConfig for a versioned datasource.

dynamodbDataSourceConfig_tableName :: Lens' DynamodbDataSourceConfig Text

The table name.

dynamodbDataSourceConfig_awsRegion :: Lens' DynamodbDataSourceConfig Text

The Amazon Web Services Region.

ElasticsearchDataSourceConfig

elasticsearchDataSourceConfig_endpoint :: Lens' ElasticsearchDataSourceConfig Text

The endpoint.

elasticsearchDataSourceConfig_awsRegion :: Lens' ElasticsearchDataSourceConfig Text

The Amazon Web Services Region.

FunctionConfiguration

functionConfiguration_responseMappingTemplate :: Lens' FunctionConfiguration (Maybe Text)

The Function response mapping template.

functionConfiguration_functionVersion :: Lens' FunctionConfiguration (Maybe Text)

The version of the request mapping template. Currently only the 2018-05-29 version of the template is supported.

functionConfiguration_syncConfig :: Lens' FunctionConfiguration (Maybe SyncConfig)

Undocumented member.

functionConfiguration_dataSourceName :: Lens' FunctionConfiguration (Maybe Text)

The name of the DataSource.

functionConfiguration_functionArn :: Lens' FunctionConfiguration (Maybe Text)

The ARN of the Function object.

functionConfiguration_name :: Lens' FunctionConfiguration (Maybe Text)

The name of the Function object.

functionConfiguration_functionId :: Lens' FunctionConfiguration (Maybe Text)

A unique ID representing the Function object.

functionConfiguration_description :: Lens' FunctionConfiguration (Maybe Text)

The Function description.

functionConfiguration_requestMappingTemplate :: Lens' FunctionConfiguration (Maybe Text)

The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

GraphqlApi

graphqlApi_wafWebAclArn :: Lens' GraphqlApi (Maybe Text)

The ARN of the WAF ACL associated with this GraphqlApi, if one exists.

graphqlApi_apiId :: Lens' GraphqlApi (Maybe Text)

The API ID.

graphqlApi_arn :: Lens' GraphqlApi (Maybe Text)

The ARN.

graphqlApi_openIDConnectConfig :: Lens' GraphqlApi (Maybe OpenIDConnectConfig)

The OpenID Connect configuration.

graphqlApi_userPoolConfig :: Lens' GraphqlApi (Maybe UserPoolConfig)

The Amazon Cognito user pool configuration.

graphqlApi_xrayEnabled :: Lens' GraphqlApi (Maybe Bool)

A flag representing whether X-Ray tracing is enabled for this GraphqlApi.

graphqlApi_name :: Lens' GraphqlApi (Maybe Text)

The API name.

graphqlApi_lambdaAuthorizerConfig :: Lens' GraphqlApi (Maybe LambdaAuthorizerConfig)

Configuration for Amazon Web Services Lambda function authorization.

graphqlApi_tags :: Lens' GraphqlApi (Maybe (HashMap Text Text))

The tags.

graphqlApi_logConfig :: Lens' GraphqlApi (Maybe LogConfig)

The Amazon CloudWatch Logs configuration.

graphqlApi_additionalAuthenticationProviders :: Lens' GraphqlApi (Maybe [AdditionalAuthenticationProvider])

A list of additional authentication providers for the GraphqlApi API.

graphqlApi_authenticationType :: Lens' GraphqlApi (Maybe AuthenticationType)

The authentication type.

graphqlApi_uris :: Lens' GraphqlApi (Maybe (HashMap Text Text))

The URIs.

HttpDataSourceConfig

httpDataSourceConfig_authorizationConfig :: Lens' HttpDataSourceConfig (Maybe AuthorizationConfig)

The authorization config in case the HTTP endpoint requires authorization.

httpDataSourceConfig_endpoint :: Lens' HttpDataSourceConfig (Maybe Text)

The HTTP URL endpoint. You can either specify the domain name or IP, and port combination, and the URL scheme must be HTTP or HTTPS. If the port is not specified, AppSync uses the default port 80 for the HTTP endpoint and port 443 for HTTPS endpoints.

LambdaAuthorizerConfig

lambdaAuthorizerConfig_identityValidationExpression :: Lens' LambdaAuthorizerConfig (Maybe Text)

A regular expression for validation of tokens before the Lambda function is called.

lambdaAuthorizerConfig_authorizerResultTtlInSeconds :: Lens' LambdaAuthorizerConfig (Maybe Natural)

The number of seconds a response should be cached for. The default is 5 minutes (300 seconds). The Lambda function can override this by returning a ttlOverride key in its response. A value of 0 disables caching of responses.

lambdaAuthorizerConfig_authorizerUri :: Lens' LambdaAuthorizerConfig Text

The ARN of the Lambda function to be called for authorization. This may be a standard Lambda ARN, a version ARN (.../v3) or alias ARN.

Note: This Lambda function must have the following resource-based policy assigned to it. When configuring Lambda authorizers in the Console, this is done for you. To do so with the Amazon Web Services CLI, run the following:

aws lambda add-permission --function-name "arn:aws:lambda:us-east-2:111122223333:function:my-function" --statement-id "appsync" --principal appsync.amazonaws.com --action lambda:InvokeFunction

LambdaConflictHandlerConfig

lambdaConflictHandlerConfig_lambdaConflictHandlerArn :: Lens' LambdaConflictHandlerConfig (Maybe Text)

The Arn for the Lambda function to use as the Conflict Handler.

LambdaDataSourceConfig

lambdaDataSourceConfig_lambdaFunctionArn :: Lens' LambdaDataSourceConfig Text

The ARN for the Lambda function.

LogConfig

logConfig_excludeVerboseContent :: Lens' LogConfig (Maybe Bool)

Set to TRUE to exclude sections that contain information such as headers, context, and evaluated mapping templates, regardless of logging level.

logConfig_fieldLogLevel :: Lens' LogConfig FieldLogLevel

The field logging level. Values can be NONE, ERROR, or ALL.

NONE: No field-level logs are captured.

ERROR: Logs the following information only for the fields that are in error:

The error section in the server response.

Field-level errors.

The generated request/response functions that got resolved for error fields.

ALL: The following information is logged for all fields in the query:

Field-level tracing information.

The generated request/response functions that got resolved for each field.

logConfig_cloudWatchLogsRoleArn :: Lens' LogConfig Text

The service role that AppSync will assume to publish to Amazon CloudWatch logs in your account.

OpenIDConnectConfig

openIDConnectConfig_clientId :: Lens' OpenIDConnectConfig (Maybe Text)

The client identifier of the Relying party at the OpenID identity provider. This identifier is typically obtained when the Relying party is registered with the OpenID identity provider. You can specify a regular expression so the AppSync can validate against multiple client identifiers at a time.

openIDConnectConfig_authTTL :: Lens' OpenIDConnectConfig (Maybe Integer)

The number of milliseconds a token is valid after being authenticated.

openIDConnectConfig_iatTTL :: Lens' OpenIDConnectConfig (Maybe Integer)

The number of milliseconds a token is valid after being issued to a user.

openIDConnectConfig_issuer :: Lens' OpenIDConnectConfig Text

The issuer for the OpenID Connect configuration. The issuer returned by discovery must exactly match the value of iss in the ID token.

OpenSearchServiceDataSourceConfig

openSearchServiceDataSourceConfig_endpoint :: Lens' OpenSearchServiceDataSourceConfig Text

The endpoint.

openSearchServiceDataSourceConfig_awsRegion :: Lens' OpenSearchServiceDataSourceConfig Text

The Amazon Web Services Region.

PipelineConfig

pipelineConfig_functions :: Lens' PipelineConfig (Maybe [Text])

A list of Function objects.

RdsHttpEndpointConfig

rdsHttpEndpointConfig_schema :: Lens' RdsHttpEndpointConfig (Maybe Text)

Logical schema name.

rdsHttpEndpointConfig_awsSecretStoreArn :: Lens' RdsHttpEndpointConfig (Maybe Text)

Amazon Web Services secret store ARN for database credentials.

rdsHttpEndpointConfig_dbClusterIdentifier :: Lens' RdsHttpEndpointConfig (Maybe Text)

Amazon RDS cluster ARN.

rdsHttpEndpointConfig_awsRegion :: Lens' RdsHttpEndpointConfig (Maybe Text)

Amazon Web Services Region for RDS HTTP endpoint.

rdsHttpEndpointConfig_databaseName :: Lens' RdsHttpEndpointConfig (Maybe Text)

Logical database name.

RelationalDatabaseDataSourceConfig

relationalDatabaseDataSourceConfig_rdsHttpEndpointConfig :: Lens' RelationalDatabaseDataSourceConfig (Maybe RdsHttpEndpointConfig)

Amazon RDS HTTP endpoint settings.

relationalDatabaseDataSourceConfig_relationalDatabaseSourceType :: Lens' RelationalDatabaseDataSourceConfig (Maybe RelationalDatabaseSourceType)

Source type for the relational database.

RDS_HTTP_ENDPOINT: The relational database source type is an Amazon RDS HTTP endpoint.

Resolver

resolver_responseMappingTemplate :: Lens' Resolver (Maybe Text)

The response mapping template.

resolver_typeName :: Lens' Resolver (Maybe Text)

The resolver type name.

resolver_kind :: Lens' Resolver (Maybe ResolverKind)

The resolver type.

UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source.

PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

resolver_syncConfig :: Lens' Resolver (Maybe SyncConfig)

The SyncConfig for a resolver attached to a versioned datasource.

resolver_dataSourceName :: Lens' Resolver (Maybe Text)

The resolver data source name.

resolver_resolverArn :: Lens' Resolver (Maybe Text)

The resolver ARN.

resolver_cachingConfig :: Lens' Resolver (Maybe CachingConfig)

The caching configuration for the resolver.

resolver_pipelineConfig :: Lens' Resolver (Maybe PipelineConfig)

The PipelineConfig.

resolver_requestMappingTemplate :: Lens' Resolver (Maybe Text)

The request mapping template.

resolver_fieldName :: Lens' Resolver (Maybe Text)

The resolver field name.

SyncConfig

syncConfig_conflictHandler :: Lens' SyncConfig (Maybe ConflictHandlerType)

The Conflict Resolution strategy to perform in the event of a conflict.

OPTIMISTIC_CONCURRENCY: Resolve conflicts by rejecting mutations when versions do not match the latest version at the server.

AUTOMERGE: Resolve conflicts with the Automerge conflict resolution strategy.

LAMBDA: Resolve conflicts with a Lambda function supplied in the LambdaConflictHandlerConfig.

syncConfig_lambdaConflictHandlerConfig :: Lens' SyncConfig (Maybe LambdaConflictHandlerConfig)

The LambdaConflictHandlerConfig when configuring LAMBDA as the Conflict Handler.

syncConfig_conflictDetection :: Lens' SyncConfig (Maybe ConflictDetectionType)

The Conflict Detection strategy to use.

VERSION: Detect conflicts based on object versions for this resolver.

NONE: Do not detect conflicts when executing this resolver.

Type

type_format :: Lens' Type (Maybe TypeDefinitionFormat)

The type format: SDL or JSON.

type_arn :: Lens' Type (Maybe Text)

The type ARN.

type_name :: Lens' Type (Maybe Text)

The type name.

type_description :: Lens' Type (Maybe Text)

The type description.

type_definition :: Lens' Type (Maybe Text)

The type definition.

UserPoolConfig

userPoolConfig_appIdClientRegex :: Lens' UserPoolConfig (Maybe Text)

A regular expression for validating the incoming Amazon Cognito user pool app client ID.

userPoolConfig_userPoolId :: Lens' UserPoolConfig Text

The user pool ID.

userPoolConfig_awsRegion :: Lens' UserPoolConfig Text

The Amazon Web Services Region in which the user pool was created.

userPoolConfig_defaultAction :: Lens' UserPoolConfig DefaultAction

The action that you want your GraphQL API to take when a request that uses Amazon Cognito user pool authentication doesn't match the Amazon Cognito user pool configuration.