Network.AWS.AppSync.ListResolversByFunction

Description

List the resolvers that are associated with a specific function.

This operation returns paginated results.

Creating a Request

data ListResolversByFunction

See: newListResolversByFunction smart constructor.

Constructors

ListResolversByFunction'

Fields

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.

maxResults :: Maybe Natural

The maximum number of results you want the request to return.

apiId :: Text

The API ID.

functionId :: Text

The Function ID.

newListResolversByFunction

Arguments

:: Text
$sel:apiId:ListResolversByFunction'

-> Text
$sel:functionId:ListResolversByFunction'

-> ListResolversByFunction

Create a value of ListResolversByFunction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListResolversByFunction', listResolversByFunction_nextToken - An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.

$sel:maxResults:ListResolversByFunction', listResolversByFunction_maxResults - The maximum number of results you want the request to return.

$sel:apiId:ListResolversByFunction', listResolversByFunction_apiId - The API ID.

$sel:functionId:ListResolversByFunction', listResolversByFunction_functionId - The Function ID.

Request Lenses

listResolversByFunction_nextToken :: Lens' ListResolversByFunction (Maybe Text)

An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.

listResolversByFunction_maxResults :: Lens' ListResolversByFunction (Maybe Natural)

The maximum number of results you want the request to return.

listResolversByFunction_apiId :: Lens' ListResolversByFunction Text

The API ID.

listResolversByFunction_functionId :: Lens' ListResolversByFunction Text

The Function ID.

Destructuring the Response

data ListResolversByFunctionResponse

See: newListResolversByFunctionResponse smart constructor.

Constructors

ListResolversByFunctionResponse'

Fields

nextToken :: Maybe Text

An identifier that can be used to return the next set of items in the list.

resolvers :: Maybe [Resolver]

The list of resolvers.

httpStatus :: Int

The response's http status code.

newListResolversByFunctionResponse

Arguments

:: Int
$sel:httpStatus:ListResolversByFunctionResponse'

-> ListResolversByFunctionResponse

Create a value of ListResolversByFunctionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListResolversByFunction', listResolversByFunctionResponse_nextToken - An identifier that can be used to return the next set of items in the list.

$sel:resolvers:ListResolversByFunctionResponse', listResolversByFunctionResponse_resolvers - The list of resolvers.

$sel:httpStatus:ListResolversByFunctionResponse', listResolversByFunctionResponse_httpStatus - The response's http status code.

Response Lenses

listResolversByFunctionResponse_nextToken :: Lens' ListResolversByFunctionResponse (Maybe Text)

An identifier that can be used to return the next set of items in the list.

listResolversByFunctionResponse_resolvers :: Lens' ListResolversByFunctionResponse (Maybe [Resolver])

The list of resolvers.

listResolversByFunctionResponse_httpStatus :: Lens' ListResolversByFunctionResponse Int

The response's http status code.