Network.AWS.AppSync.Types.ApiCache

Description

Documentation

data ApiCache

The ApiCache object.

See: newApiCache smart constructor.

Constructors

ApiCache'

Fields

status :: Maybe ApiCacheStatus

The cache instance status.

AVAILABLE: The instance is available for use.

CREATING: The instance is currently creating.

DELETING: The instance is currently deleting.

MODIFYING: The instance is currently modifying.

FAILED: The instance has failed creation.

atRestEncryptionEnabled :: Maybe Bool

At rest encryption flag for cache. This setting cannot be updated after creation.

ttl :: Maybe Integer

TTL in seconds for cache entries.

Valid values are between 1 and 3600 seconds.

type' :: Maybe ApiCacheType

The cache instance type. Valid values are

SMALL

MEDIUM

LARGE

XLARGE

LARGE_2X

LARGE_4X

LARGE_8X (not available in all regions)

LARGE_12X

Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used.

The following legacy instance types are available, but their use is discouraged:

T2_SMALL: A t2.small instance type.

T2_MEDIUM: A t2.medium instance type.

R4_LARGE: A r4.large instance type.

R4_XLARGE: A r4.xlarge instance type.

R4_2XLARGE: A r4.2xlarge instance type.

R4_4XLARGE: A r4.4xlarge instance type.

R4_8XLARGE: A r4.8xlarge instance type.

transitEncryptionEnabled :: Maybe Bool

Transit encryption flag when connecting to cache. This setting cannot be updated after creation.

apiCachingBehavior :: Maybe ApiCachingBehavior

Caching behavior.

FULL_REQUEST_CACHING: All requests are fully cached.

PER_RESOLVER_CACHING: Individual resolvers that you specify are cached.

newApiCache :: ApiCache

Create a value of ApiCache with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ApiCache', apiCache_status - The cache instance status.

AVAILABLE: The instance is available for use.

CREATING: The instance is currently creating.

DELETING: The instance is currently deleting.

MODIFYING: The instance is currently modifying.

FAILED: The instance has failed creation.

$sel:atRestEncryptionEnabled:ApiCache', apiCache_atRestEncryptionEnabled - At rest encryption flag for cache. This setting cannot be updated after creation.

$sel:ttl:ApiCache', apiCache_ttl - TTL in seconds for cache entries.

Valid values are between 1 and 3600 seconds.

$sel:type':ApiCache', apiCache_type - The cache instance type. Valid values are

SMALL

MEDIUM

LARGE

XLARGE

LARGE_2X

LARGE_4X

LARGE_8X (not available in all regions)

LARGE_12X

Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used.

The following legacy instance types are available, but their use is discouraged:

T2_SMALL: A t2.small instance type.

T2_MEDIUM: A t2.medium instance type.

R4_LARGE: A r4.large instance type.

R4_XLARGE: A r4.xlarge instance type.

R4_2XLARGE: A r4.2xlarge instance type.

R4_4XLARGE: A r4.4xlarge instance type.

R4_8XLARGE: A r4.8xlarge instance type.

$sel:transitEncryptionEnabled:ApiCache', apiCache_transitEncryptionEnabled - Transit encryption flag when connecting to cache. This setting cannot be updated after creation.

$sel:apiCachingBehavior:ApiCache', apiCache_apiCachingBehavior - Caching behavior.

FULL_REQUEST_CACHING: All requests are fully cached.

PER_RESOLVER_CACHING: Individual resolvers that you specify are cached.

apiCache_status :: Lens' ApiCache (Maybe ApiCacheStatus)

The cache instance status.

AVAILABLE: The instance is available for use.

CREATING: The instance is currently creating.

DELETING: The instance is currently deleting.

MODIFYING: The instance is currently modifying.

FAILED: The instance has failed creation.

apiCache_atRestEncryptionEnabled :: Lens' ApiCache (Maybe Bool)

At rest encryption flag for cache. This setting cannot be updated after creation.

apiCache_ttl :: Lens' ApiCache (Maybe Integer)

TTL in seconds for cache entries.

Valid values are between 1 and 3600 seconds.

apiCache_type :: Lens' ApiCache (Maybe ApiCacheType)

The cache instance type. Valid values are

SMALL

MEDIUM

LARGE

XLARGE

LARGE_2X

LARGE_4X

LARGE_8X (not available in all regions)

LARGE_12X

Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used.

The following legacy instance types are available, but their use is discouraged:

T2_SMALL: A t2.small instance type.

T2_MEDIUM: A t2.medium instance type.

R4_LARGE: A r4.large instance type.

R4_XLARGE: A r4.xlarge instance type.

R4_2XLARGE: A r4.2xlarge instance type.

R4_4XLARGE: A r4.4xlarge instance type.

R4_8XLARGE: A r4.8xlarge instance type.

apiCache_transitEncryptionEnabled :: Lens' ApiCache (Maybe Bool)

Transit encryption flag when connecting to cache. This setting cannot be updated after creation.

apiCache_apiCachingBehavior :: Lens' ApiCache (Maybe ApiCachingBehavior)

Caching behavior.

FULL_REQUEST_CACHING: All requests are fully cached.

PER_RESOLVER_CACHING: Individual resolvers that you specify are cached.