Network.AWS.AppSync.Types.DataSource

Description

Documentation

data DataSource

Describes a data source.

See: newDataSource smart constructor.

Constructors

DataSource'

Fields

serviceRoleArn :: Maybe Text

The Identity and Access Management service role ARN for the data source. The system assumes this role when accessing the data source.

relationalDatabaseConfig :: Maybe RelationalDatabaseDataSourceConfig

Relational database settings.

elasticsearchConfig :: Maybe ElasticsearchDataSourceConfig

Amazon OpenSearch Service settings.

lambdaConfig :: Maybe LambdaDataSourceConfig

Amazon Web Services Lambda settings.

name :: Maybe Text

The name of the data source.

dynamodbConfig :: Maybe DynamodbDataSourceConfig

Amazon DynamoDB settings.

description :: Maybe Text

The description of the data source.

dataSourceArn :: Maybe Text

The data source ARN.

type' :: Maybe DataSourceType

The type of the data source.

AWS_LAMBDA: The data source is an Amazon Web Services Lambda function.

AMAZON_DYNAMODB: The data source is an Amazon DynamoDB table.

AMAZON_ELASTICSEARCH: The data source is an Amazon OpenSearch Service domain.

AMAZON_OPENSEARCH_SERVICE: The data source is an Amazon OpenSearch Service domain.

NONE: There is no data source. This type is used when you wish to invoke a GraphQL operation without connecting to a data source, such as performing data transformation with resolvers or triggering a subscription to be invoked from a mutation.

HTTP: The data source is an HTTP endpoint.

RELATIONAL_DATABASE: The data source is a relational database.

openSearchServiceConfig :: Maybe OpenSearchServiceDataSourceConfig

Amazon OpenSearch Service settings.

httpConfig :: Maybe HttpDataSourceConfig

HTTP endpoint settings.

newDataSource :: DataSource

Create a value of DataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceRoleArn:DataSource', dataSource_serviceRoleArn - The Identity and Access Management service role ARN for the data source. The system assumes this role when accessing the data source.

$sel:relationalDatabaseConfig:DataSource', dataSource_relationalDatabaseConfig - Relational database settings.

$sel:elasticsearchConfig:DataSource', dataSource_elasticsearchConfig - Amazon OpenSearch Service settings.

$sel:lambdaConfig:DataSource', dataSource_lambdaConfig - Amazon Web Services Lambda settings.

$sel:name:DataSource', dataSource_name - The name of the data source.

$sel:dynamodbConfig:DataSource', dataSource_dynamodbConfig - Amazon DynamoDB settings.

$sel:description:DataSource', dataSource_description - The description of the data source.

$sel:dataSourceArn:DataSource', dataSource_dataSourceArn - The data source ARN.

$sel:type':DataSource', dataSource_type - The type of the data source.

AWS_LAMBDA: The data source is an Amazon Web Services Lambda function.

AMAZON_DYNAMODB: The data source is an Amazon DynamoDB table.

AMAZON_ELASTICSEARCH: The data source is an Amazon OpenSearch Service domain.

AMAZON_OPENSEARCH_SERVICE: The data source is an Amazon OpenSearch Service domain.

NONE: There is no data source. This type is used when you wish to invoke a GraphQL operation without connecting to a data source, such as performing data transformation with resolvers or triggering a subscription to be invoked from a mutation.

HTTP: The data source is an HTTP endpoint.

RELATIONAL_DATABASE: The data source is a relational database.

$sel:openSearchServiceConfig:DataSource', dataSource_openSearchServiceConfig - Amazon OpenSearch Service settings.

$sel:httpConfig:DataSource', dataSource_httpConfig - HTTP endpoint settings.

dataSource_serviceRoleArn :: Lens' DataSource (Maybe Text)

The Identity and Access Management service role ARN for the data source. The system assumes this role when accessing the data source.

dataSource_relationalDatabaseConfig :: Lens' DataSource (Maybe RelationalDatabaseDataSourceConfig)

Relational database settings.

dataSource_elasticsearchConfig :: Lens' DataSource (Maybe ElasticsearchDataSourceConfig)

Amazon OpenSearch Service settings.

dataSource_lambdaConfig :: Lens' DataSource (Maybe LambdaDataSourceConfig)

Amazon Web Services Lambda settings.

dataSource_name :: Lens' DataSource (Maybe Text)

The name of the data source.

dataSource_dynamodbConfig :: Lens' DataSource (Maybe DynamodbDataSourceConfig)

Amazon DynamoDB settings.

dataSource_description :: Lens' DataSource (Maybe Text)

The description of the data source.

dataSource_dataSourceArn :: Lens' DataSource (Maybe Text)

The data source ARN.

dataSource_type :: Lens' DataSource (Maybe DataSourceType)

The type of the data source.

AWS_LAMBDA: The data source is an Amazon Web Services Lambda function.

AMAZON_DYNAMODB: The data source is an Amazon DynamoDB table.

AMAZON_ELASTICSEARCH: The data source is an Amazon OpenSearch Service domain.

AMAZON_OPENSEARCH_SERVICE: The data source is an Amazon OpenSearch Service domain.

NONE: There is no data source. This type is used when you wish to invoke a GraphQL operation without connecting to a data source, such as performing data transformation with resolvers or triggering a subscription to be invoked from a mutation.

HTTP: The data source is an HTTP endpoint.

RELATIONAL_DATABASE: The data source is a relational database.

dataSource_openSearchServiceConfig :: Lens' DataSource (Maybe OpenSearchServiceDataSourceConfig)

Amazon OpenSearch Service settings.

dataSource_httpConfig :: Lens' DataSource (Maybe HttpDataSourceConfig)

HTTP endpoint settings.