Network.AWS.AppSync.Types.GraphqlApi

Description

Documentation

data GraphqlApi

Describes a GraphQL API.

See: newGraphqlApi smart constructor.

Constructors

GraphqlApi'

Fields

wafWebAclArn :: Maybe Text

The ARN of the WAF ACL associated with this GraphqlApi, if one exists.

apiId :: Maybe Text

The API ID.

arn :: Maybe Text

The ARN.

openIDConnectConfig :: Maybe OpenIDConnectConfig

The OpenID Connect configuration.

userPoolConfig :: Maybe UserPoolConfig

The Amazon Cognito user pool configuration.

xrayEnabled :: Maybe Bool

A flag representing whether X-Ray tracing is enabled for this GraphqlApi.

name :: Maybe Text

The API name.

lambdaAuthorizerConfig :: Maybe LambdaAuthorizerConfig

Configuration for Amazon Web Services Lambda function authorization.

tags :: Maybe (HashMap Text Text)

The tags.

logConfig :: Maybe LogConfig

The Amazon CloudWatch Logs configuration.

additionalAuthenticationProviders :: Maybe [AdditionalAuthenticationProvider]

A list of additional authentication providers for the GraphqlApi API.

authenticationType :: Maybe AuthenticationType

The authentication type.

uris :: Maybe (HashMap Text Text)

The URIs.

newGraphqlApi :: GraphqlApi

Create a value of GraphqlApi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:wafWebAclArn:GraphqlApi', graphqlApi_wafWebAclArn - The ARN of the WAF ACL associated with this GraphqlApi, if one exists.

$sel:apiId:GraphqlApi', graphqlApi_apiId - The API ID.

$sel:arn:GraphqlApi', graphqlApi_arn - The ARN.

$sel:openIDConnectConfig:GraphqlApi', graphqlApi_openIDConnectConfig - The OpenID Connect configuration.

$sel:userPoolConfig:GraphqlApi', graphqlApi_userPoolConfig - The Amazon Cognito user pool configuration.

$sel:xrayEnabled:GraphqlApi', graphqlApi_xrayEnabled - A flag representing whether X-Ray tracing is enabled for this GraphqlApi.

$sel:name:GraphqlApi', graphqlApi_name - The API name.

$sel:lambdaAuthorizerConfig:GraphqlApi', graphqlApi_lambdaAuthorizerConfig - Configuration for Amazon Web Services Lambda function authorization.

$sel:tags:GraphqlApi', graphqlApi_tags - The tags.

$sel:logConfig:GraphqlApi', graphqlApi_logConfig - The Amazon CloudWatch Logs configuration.

$sel:additionalAuthenticationProviders:GraphqlApi', graphqlApi_additionalAuthenticationProviders - A list of additional authentication providers for the GraphqlApi API.

$sel:authenticationType:GraphqlApi', graphqlApi_authenticationType - The authentication type.

$sel:uris:GraphqlApi', graphqlApi_uris - The URIs.

graphqlApi_wafWebAclArn :: Lens' GraphqlApi (Maybe Text)

The ARN of the WAF ACL associated with this GraphqlApi, if one exists.

graphqlApi_apiId :: Lens' GraphqlApi (Maybe Text)

The API ID.

graphqlApi_arn :: Lens' GraphqlApi (Maybe Text)

The ARN.

graphqlApi_openIDConnectConfig :: Lens' GraphqlApi (Maybe OpenIDConnectConfig)

The OpenID Connect configuration.

graphqlApi_userPoolConfig :: Lens' GraphqlApi (Maybe UserPoolConfig)

The Amazon Cognito user pool configuration.

graphqlApi_xrayEnabled :: Lens' GraphqlApi (Maybe Bool)

A flag representing whether X-Ray tracing is enabled for this GraphqlApi.

graphqlApi_name :: Lens' GraphqlApi (Maybe Text)

The API name.

graphqlApi_lambdaAuthorizerConfig :: Lens' GraphqlApi (Maybe LambdaAuthorizerConfig)

Configuration for Amazon Web Services Lambda function authorization.

graphqlApi_tags :: Lens' GraphqlApi (Maybe (HashMap Text Text))

The tags.

graphqlApi_logConfig :: Lens' GraphqlApi (Maybe LogConfig)

The Amazon CloudWatch Logs configuration.

graphqlApi_additionalAuthenticationProviders :: Lens' GraphqlApi (Maybe [AdditionalAuthenticationProvider])

A list of additional authentication providers for the GraphqlApi API.

graphqlApi_authenticationType :: Lens' GraphqlApi (Maybe AuthenticationType)

The authentication type.

graphqlApi_uris :: Lens' GraphqlApi (Maybe (HashMap Text Text))

The URIs.