Network.AWS.AppSync.Types.HttpDataSourceConfig

Description

Documentation

data HttpDataSourceConfig

Describes an HTTP data source configuration.

See: newHttpDataSourceConfig smart constructor.

Constructors

HttpDataSourceConfig'

Fields

authorizationConfig :: Maybe AuthorizationConfig

The authorization config in case the HTTP endpoint requires authorization.

endpoint :: Maybe Text

The HTTP URL endpoint. You can either specify the domain name or IP, and port combination, and the URL scheme must be HTTP or HTTPS. If the port is not specified, AppSync uses the default port 80 for the HTTP endpoint and port 443 for HTTPS endpoints.

newHttpDataSourceConfig :: HttpDataSourceConfig

Create a value of HttpDataSourceConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizationConfig:HttpDataSourceConfig', httpDataSourceConfig_authorizationConfig - The authorization config in case the HTTP endpoint requires authorization.

$sel:endpoint:HttpDataSourceConfig', httpDataSourceConfig_endpoint - The HTTP URL endpoint. You can either specify the domain name or IP, and port combination, and the URL scheme must be HTTP or HTTPS. If the port is not specified, AppSync uses the default port 80 for the HTTP endpoint and port 443 for HTTPS endpoints.

httpDataSourceConfig_authorizationConfig :: Lens' HttpDataSourceConfig (Maybe AuthorizationConfig)

The authorization config in case the HTTP endpoint requires authorization.

httpDataSourceConfig_endpoint :: Lens' HttpDataSourceConfig (Maybe Text)

The HTTP URL endpoint. You can either specify the domain name or IP, and port combination, and the URL scheme must be HTTP or HTTPS. If the port is not specified, AppSync uses the default port 80 for the HTTP endpoint and port 443 for HTTPS endpoints.