Network.AWS.AppSync.Types.LogConfig

Description

Documentation

data LogConfig

The CloudWatch Logs configuration.

See: newLogConfig smart constructor.

Constructors

LogConfig'

Fields

excludeVerboseContent :: Maybe Bool

Set to TRUE to exclude sections that contain information such as headers, context, and evaluated mapping templates, regardless of logging level.

fieldLogLevel :: FieldLogLevel

The field logging level. Values can be NONE, ERROR, or ALL.

NONE: No field-level logs are captured.

ERROR: Logs the following information only for the fields that are in error:

The error section in the server response.

Field-level errors.

The generated request/response functions that got resolved for error fields.

ALL: The following information is logged for all fields in the query:

Field-level tracing information.

The generated request/response functions that got resolved for each field.

cloudWatchLogsRoleArn :: Text

The service role that AppSync will assume to publish to Amazon CloudWatch logs in your account.

newLogConfig

Arguments

:: FieldLogLevel
$sel:fieldLogLevel:LogConfig'

-> Text
$sel:cloudWatchLogsRoleArn:LogConfig'

-> LogConfig

Create a value of LogConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:excludeVerboseContent:LogConfig', logConfig_excludeVerboseContent - Set to TRUE to exclude sections that contain information such as headers, context, and evaluated mapping templates, regardless of logging level.

$sel:fieldLogLevel:LogConfig', logConfig_fieldLogLevel - The field logging level. Values can be NONE, ERROR, or ALL.

NONE: No field-level logs are captured.

ERROR: Logs the following information only for the fields that are in error:

The error section in the server response.

Field-level errors.

The generated request/response functions that got resolved for error fields.

ALL: The following information is logged for all fields in the query:

Field-level tracing information.

The generated request/response functions that got resolved for each field.

$sel:cloudWatchLogsRoleArn:LogConfig', logConfig_cloudWatchLogsRoleArn - The service role that AppSync will assume to publish to Amazon CloudWatch logs in your account.

logConfig_excludeVerboseContent :: Lens' LogConfig (Maybe Bool)

Set to TRUE to exclude sections that contain information such as headers, context, and evaluated mapping templates, regardless of logging level.

logConfig_fieldLogLevel :: Lens' LogConfig FieldLogLevel

The field logging level. Values can be NONE, ERROR, or ALL.

NONE: No field-level logs are captured.

ERROR: Logs the following information only for the fields that are in error:

The error section in the server response.

Field-level errors.

The generated request/response functions that got resolved for error fields.

ALL: The following information is logged for all fields in the query:

Field-level tracing information.

The generated request/response functions that got resolved for each field.

logConfig_cloudWatchLogsRoleArn :: Lens' LogConfig Text

The service role that AppSync will assume to publish to Amazon CloudWatch logs in your account.