Network.AWS.AppSync.Types.OpenIDConnectConfig

Description

Documentation

data OpenIDConnectConfig

Describes an OpenID Connect configuration.

See: newOpenIDConnectConfig smart constructor.

Constructors

OpenIDConnectConfig'

Fields

clientId :: Maybe Text

The client identifier of the Relying party at the OpenID identity provider. This identifier is typically obtained when the Relying party is registered with the OpenID identity provider. You can specify a regular expression so the AppSync can validate against multiple client identifiers at a time.

authTTL :: Maybe Integer

The number of milliseconds a token is valid after being authenticated.

iatTTL :: Maybe Integer

The number of milliseconds a token is valid after being issued to a user.

issuer :: Text

The issuer for the OpenID Connect configuration. The issuer returned by discovery must exactly match the value of iss in the ID token.

newOpenIDConnectConfig

Arguments

:: Text
$sel:issuer:OpenIDConnectConfig'

-> OpenIDConnectConfig

Create a value of OpenIDConnectConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:OpenIDConnectConfig', openIDConnectConfig_clientId - The client identifier of the Relying party at the OpenID identity provider. This identifier is typically obtained when the Relying party is registered with the OpenID identity provider. You can specify a regular expression so the AppSync can validate against multiple client identifiers at a time.

$sel:authTTL:OpenIDConnectConfig', openIDConnectConfig_authTTL - The number of milliseconds a token is valid after being authenticated.

$sel:iatTTL:OpenIDConnectConfig', openIDConnectConfig_iatTTL - The number of milliseconds a token is valid after being issued to a user.

$sel:issuer:OpenIDConnectConfig', openIDConnectConfig_issuer - The issuer for the OpenID Connect configuration. The issuer returned by discovery must exactly match the value of iss in the ID token.

openIDConnectConfig_clientId :: Lens' OpenIDConnectConfig (Maybe Text)

The client identifier of the Relying party at the OpenID identity provider. This identifier is typically obtained when the Relying party is registered with the OpenID identity provider. You can specify a regular expression so the AppSync can validate against multiple client identifiers at a time.

openIDConnectConfig_authTTL :: Lens' OpenIDConnectConfig (Maybe Integer)

The number of milliseconds a token is valid after being authenticated.

openIDConnectConfig_iatTTL :: Lens' OpenIDConnectConfig (Maybe Integer)

The number of milliseconds a token is valid after being issued to a user.

openIDConnectConfig_issuer :: Lens' OpenIDConnectConfig Text

The issuer for the OpenID Connect configuration. The issuer returned by discovery must exactly match the value of iss in the ID token.