Network.AWS.AppSync.Types.SyncConfig

Description

Documentation

data SyncConfig

Describes a Sync configuration for a resolver.

Contains information on which Conflict Detection as well as Resolution strategy should be performed when the resolver is invoked.

See: newSyncConfig smart constructor.

Constructors

SyncConfig'

Fields

conflictHandler :: Maybe ConflictHandlerType

The Conflict Resolution strategy to perform in the event of a conflict.

OPTIMISTIC_CONCURRENCY: Resolve conflicts by rejecting mutations when versions do not match the latest version at the server.

AUTOMERGE: Resolve conflicts with the Automerge conflict resolution strategy.

LAMBDA: Resolve conflicts with a Lambda function supplied in the LambdaConflictHandlerConfig.

lambdaConflictHandlerConfig :: Maybe LambdaConflictHandlerConfig

The LambdaConflictHandlerConfig when configuring LAMBDA as the Conflict Handler.

conflictDetection :: Maybe ConflictDetectionType

The Conflict Detection strategy to use.

VERSION: Detect conflicts based on object versions for this resolver.

NONE: Do not detect conflicts when executing this resolver.

newSyncConfig :: SyncConfig

Create a value of SyncConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conflictHandler:SyncConfig', syncConfig_conflictHandler - The Conflict Resolution strategy to perform in the event of a conflict.

OPTIMISTIC_CONCURRENCY: Resolve conflicts by rejecting mutations when versions do not match the latest version at the server.

AUTOMERGE: Resolve conflicts with the Automerge conflict resolution strategy.

LAMBDA: Resolve conflicts with a Lambda function supplied in the LambdaConflictHandlerConfig.

$sel:lambdaConflictHandlerConfig:SyncConfig', syncConfig_lambdaConflictHandlerConfig - The LambdaConflictHandlerConfig when configuring LAMBDA as the Conflict Handler.

$sel:conflictDetection:SyncConfig', syncConfig_conflictDetection - The Conflict Detection strategy to use.

VERSION: Detect conflicts based on object versions for this resolver.

NONE: Do not detect conflicts when executing this resolver.

syncConfig_conflictHandler :: Lens' SyncConfig (Maybe ConflictHandlerType)

The Conflict Resolution strategy to perform in the event of a conflict.

OPTIMISTIC_CONCURRENCY: Resolve conflicts by rejecting mutations when versions do not match the latest version at the server.

AUTOMERGE: Resolve conflicts with the Automerge conflict resolution strategy.

LAMBDA: Resolve conflicts with a Lambda function supplied in the LambdaConflictHandlerConfig.

syncConfig_lambdaConflictHandlerConfig :: Lens' SyncConfig (Maybe LambdaConflictHandlerConfig)

The LambdaConflictHandlerConfig when configuring LAMBDA as the Conflict Handler.

syncConfig_conflictDetection :: Lens' SyncConfig (Maybe ConflictDetectionType)

The Conflict Detection strategy to use.

VERSION: Detect conflicts based on object versions for this resolver.

NONE: Do not detect conflicts when executing this resolver.