Network.AWS.AppSync.UpdateApiCache

Description

Updates the cache for the GraphQL API.

Creating a Request

data UpdateApiCache

Represents the input of a UpdateApiCache operation.

See: newUpdateApiCache smart constructor.

Constructors

UpdateApiCache'

Fields

apiId :: Text

The GraphQL API Id.

ttl :: Integer

TTL in seconds for cache entries.

Valid values are between 1 and 3600 seconds.

apiCachingBehavior :: ApiCachingBehavior

Caching behavior.

FULL_REQUEST_CACHING: All requests are fully cached.

PER_RESOLVER_CACHING: Individual resolvers that you specify are cached.

type' :: ApiCacheType

The cache instance type. Valid values are

SMALL

MEDIUM

LARGE

XLARGE

LARGE_2X

LARGE_4X

LARGE_8X (not available in all regions)

LARGE_12X

Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used.

The following legacy instance types are available, but their use is discouraged:

T2_SMALL: A t2.small instance type.

T2_MEDIUM: A t2.medium instance type.

R4_LARGE: A r4.large instance type.

R4_XLARGE: A r4.xlarge instance type.

R4_2XLARGE: A r4.2xlarge instance type.

R4_4XLARGE: A r4.4xlarge instance type.

R4_8XLARGE: A r4.8xlarge instance type.

newUpdateApiCache

Arguments

:: Text
$sel:apiId:UpdateApiCache'

-> Integer
$sel:ttl:UpdateApiCache'

-> ApiCachingBehavior
$sel:apiCachingBehavior:UpdateApiCache'

-> ApiCacheType
$sel:type':UpdateApiCache'

-> UpdateApiCache

Create a value of UpdateApiCache with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiId:UpdateApiCache', updateApiCache_apiId - The GraphQL API Id.

$sel:ttl:UpdateApiCache', updateApiCache_ttl - TTL in seconds for cache entries.

Valid values are between 1 and 3600 seconds.

$sel:apiCachingBehavior:UpdateApiCache', updateApiCache_apiCachingBehavior - Caching behavior.

FULL_REQUEST_CACHING: All requests are fully cached.

PER_RESOLVER_CACHING: Individual resolvers that you specify are cached.

$sel:type':UpdateApiCache', updateApiCache_type - The cache instance type. Valid values are

SMALL

MEDIUM

LARGE

XLARGE

LARGE_2X

LARGE_4X

LARGE_8X (not available in all regions)

LARGE_12X

Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used.

The following legacy instance types are available, but their use is discouraged:

T2_SMALL: A t2.small instance type.

T2_MEDIUM: A t2.medium instance type.

R4_LARGE: A r4.large instance type.

R4_XLARGE: A r4.xlarge instance type.

R4_2XLARGE: A r4.2xlarge instance type.

R4_4XLARGE: A r4.4xlarge instance type.

R4_8XLARGE: A r4.8xlarge instance type.

Request Lenses

updateApiCache_apiId :: Lens' UpdateApiCache Text

The GraphQL API Id.

updateApiCache_ttl :: Lens' UpdateApiCache Integer

TTL in seconds for cache entries.

Valid values are between 1 and 3600 seconds.

updateApiCache_apiCachingBehavior :: Lens' UpdateApiCache ApiCachingBehavior

Caching behavior.

FULL_REQUEST_CACHING: All requests are fully cached.

PER_RESOLVER_CACHING: Individual resolvers that you specify are cached.

updateApiCache_type :: Lens' UpdateApiCache ApiCacheType

The cache instance type. Valid values are

SMALL

MEDIUM

LARGE

XLARGE

LARGE_2X

LARGE_4X

LARGE_8X (not available in all regions)

LARGE_12X

Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used.

The following legacy instance types are available, but their use is discouraged:

T2_SMALL: A t2.small instance type.

T2_MEDIUM: A t2.medium instance type.

R4_LARGE: A r4.large instance type.

R4_XLARGE: A r4.xlarge instance type.

R4_2XLARGE: A r4.2xlarge instance type.

R4_4XLARGE: A r4.4xlarge instance type.

R4_8XLARGE: A r4.8xlarge instance type.

Destructuring the Response

data UpdateApiCacheResponse

Represents the output of a UpdateApiCache operation.

See: newUpdateApiCacheResponse smart constructor.

Constructors

UpdateApiCacheResponse'

Fields

apiCache :: Maybe ApiCache

The ApiCache object.

httpStatus :: Int

The response's http status code.

newUpdateApiCacheResponse

Arguments

:: Int
$sel:httpStatus:UpdateApiCacheResponse'

-> UpdateApiCacheResponse

Create a value of UpdateApiCacheResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiCache:UpdateApiCacheResponse', updateApiCacheResponse_apiCache - The ApiCache object.

$sel:httpStatus:UpdateApiCacheResponse', updateApiCacheResponse_httpStatus - The response's http status code.

Response Lenses

updateApiCacheResponse_apiCache :: Lens' UpdateApiCacheResponse (Maybe ApiCache)

The ApiCache object.

updateApiCacheResponse_httpStatus :: Lens' UpdateApiCacheResponse Int

The response's http status code.