Network.AWS.AppSync.UpdateDataSource

Description

Updates a DataSource object.

Creating a Request

data UpdateDataSource

See: newUpdateDataSource smart constructor.

Constructors

UpdateDataSource'

Fields

serviceRoleArn :: Maybe Text

The new service role ARN for the data source.

relationalDatabaseConfig :: Maybe RelationalDatabaseDataSourceConfig

The new relational database configuration.

elasticsearchConfig :: Maybe ElasticsearchDataSourceConfig

The new OpenSearch configuration.

As of September 2021, Amazon Elasticsearch service is Amazon OpenSearch Service. This configuration is deprecated. Instead, use UpdateDataSourceRequest$openSearchServiceConfig to update an OpenSearch data source.

lambdaConfig :: Maybe LambdaDataSourceConfig

The new Amazon Web Services Lambda configuration.

dynamodbConfig :: Maybe DynamodbDataSourceConfig

The new Amazon DynamoDB configuration.

description :: Maybe Text

The new description for the data source.

openSearchServiceConfig :: Maybe OpenSearchServiceDataSourceConfig

The new OpenSearch configuration.

httpConfig :: Maybe HttpDataSourceConfig

The new HTTP endpoint configuration.

apiId :: Text

The API ID.

name :: Text

The new name for the data source.

type' :: DataSourceType

The new data source type.

newUpdateDataSource

Arguments

:: Text
$sel:apiId:UpdateDataSource'

-> Text
$sel:name:UpdateDataSource'

-> DataSourceType
$sel:type':UpdateDataSource'

-> UpdateDataSource

Create a value of UpdateDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceRoleArn:UpdateDataSource', updateDataSource_serviceRoleArn - The new service role ARN for the data source.

$sel:relationalDatabaseConfig:UpdateDataSource', updateDataSource_relationalDatabaseConfig - The new relational database configuration.

$sel:elasticsearchConfig:UpdateDataSource', updateDataSource_elasticsearchConfig - The new OpenSearch configuration.

As of September 2021, Amazon Elasticsearch service is Amazon OpenSearch Service. This configuration is deprecated. Instead, use UpdateDataSourceRequest$openSearchServiceConfig to update an OpenSearch data source.

$sel:lambdaConfig:UpdateDataSource', updateDataSource_lambdaConfig - The new Amazon Web Services Lambda configuration.

$sel:dynamodbConfig:UpdateDataSource', updateDataSource_dynamodbConfig - The new Amazon DynamoDB configuration.

$sel:description:UpdateDataSource', updateDataSource_description - The new description for the data source.

$sel:openSearchServiceConfig:UpdateDataSource', updateDataSource_openSearchServiceConfig - The new OpenSearch configuration.

$sel:httpConfig:UpdateDataSource', updateDataSource_httpConfig - The new HTTP endpoint configuration.

$sel:apiId:UpdateDataSource', updateDataSource_apiId - The API ID.

$sel:name:UpdateDataSource', updateDataSource_name - The new name for the data source.

$sel:type':UpdateDataSource', updateDataSource_type - The new data source type.

Request Lenses

updateDataSource_serviceRoleArn :: Lens' UpdateDataSource (Maybe Text)

The new service role ARN for the data source.

updateDataSource_relationalDatabaseConfig :: Lens' UpdateDataSource (Maybe RelationalDatabaseDataSourceConfig)

The new relational database configuration.

updateDataSource_elasticsearchConfig :: Lens' UpdateDataSource (Maybe ElasticsearchDataSourceConfig)

The new OpenSearch configuration.

As of September 2021, Amazon Elasticsearch service is Amazon OpenSearch Service. This configuration is deprecated. Instead, use UpdateDataSourceRequest$openSearchServiceConfig to update an OpenSearch data source.

updateDataSource_lambdaConfig :: Lens' UpdateDataSource (Maybe LambdaDataSourceConfig)

The new Amazon Web Services Lambda configuration.

updateDataSource_dynamodbConfig :: Lens' UpdateDataSource (Maybe DynamodbDataSourceConfig)

The new Amazon DynamoDB configuration.

updateDataSource_description :: Lens' UpdateDataSource (Maybe Text)

The new description for the data source.

updateDataSource_openSearchServiceConfig :: Lens' UpdateDataSource (Maybe OpenSearchServiceDataSourceConfig)

The new OpenSearch configuration.

updateDataSource_httpConfig :: Lens' UpdateDataSource (Maybe HttpDataSourceConfig)

The new HTTP endpoint configuration.

updateDataSource_apiId :: Lens' UpdateDataSource Text

The API ID.

updateDataSource_name :: Lens' UpdateDataSource Text

The new name for the data source.

updateDataSource_type :: Lens' UpdateDataSource DataSourceType

The new data source type.

Destructuring the Response

data UpdateDataSourceResponse

See: newUpdateDataSourceResponse smart constructor.

Constructors

UpdateDataSourceResponse'

Fields

dataSource :: Maybe DataSource

The updated DataSource object.

httpStatus :: Int

The response's http status code.

newUpdateDataSourceResponse

Arguments

:: Int
$sel:httpStatus:UpdateDataSourceResponse'

-> UpdateDataSourceResponse

Create a value of UpdateDataSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSource:UpdateDataSourceResponse', updateDataSourceResponse_dataSource - The updated DataSource object.

$sel:httpStatus:UpdateDataSourceResponse', updateDataSourceResponse_httpStatus - The response's http status code.

Response Lenses

updateDataSourceResponse_dataSource :: Lens' UpdateDataSourceResponse (Maybe DataSource)

The updated DataSource object.

updateDataSourceResponse_httpStatus :: Lens' UpdateDataSourceResponse Int

The response's http status code.