Network.AWS.AppSync.UpdateFunction

Description

Updates a Function object.

Creating a Request

data UpdateFunction

See: newUpdateFunction smart constructor.

Constructors

UpdateFunction'

Fields

requestMappingTemplate :: Maybe Text

The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

responseMappingTemplate :: Maybe Text

The Function request mapping template.

syncConfig :: Maybe SyncConfig

description :: Maybe Text

The Function description.

apiId :: Text

The GraphQL API ID.

name :: Text

The Function name.

functionId :: Text

The function ID.

dataSourceName :: Text

The Function DataSource name.

functionVersion :: Text

The version of the request mapping template. Currently the supported value is 2018-05-29.

newUpdateFunction

Arguments

:: Text
$sel:apiId:UpdateFunction'

-> Text
$sel:name:UpdateFunction'

-> Text
$sel:functionId:UpdateFunction'

-> Text
$sel:dataSourceName:UpdateFunction'

-> Text
$sel:functionVersion:UpdateFunction'

-> UpdateFunction

Create a value of UpdateFunction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestMappingTemplate:UpdateFunction', updateFunction_requestMappingTemplate - The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

$sel:responseMappingTemplate:UpdateFunction', updateFunction_responseMappingTemplate - The Function request mapping template.

$sel:syncConfig:UpdateFunction', updateFunction_syncConfig - Undocumented member.

$sel:description:UpdateFunction', updateFunction_description - The Function description.

$sel:apiId:UpdateFunction', updateFunction_apiId - The GraphQL API ID.

$sel:name:UpdateFunction', updateFunction_name - The Function name.

$sel:functionId:UpdateFunction', updateFunction_functionId - The function ID.

$sel:dataSourceName:UpdateFunction', updateFunction_dataSourceName - The Function DataSource name.

$sel:functionVersion:UpdateFunction', updateFunction_functionVersion - The version of the request mapping template. Currently the supported value is 2018-05-29.

Request Lenses

updateFunction_requestMappingTemplate :: Lens' UpdateFunction (Maybe Text)

The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

updateFunction_responseMappingTemplate :: Lens' UpdateFunction (Maybe Text)

The Function request mapping template.

updateFunction_syncConfig :: Lens' UpdateFunction (Maybe SyncConfig)

Undocumented member.

updateFunction_description :: Lens' UpdateFunction (Maybe Text)

The Function description.

updateFunction_apiId :: Lens' UpdateFunction Text

The GraphQL API ID.

updateFunction_name :: Lens' UpdateFunction Text

The Function name.

updateFunction_functionId :: Lens' UpdateFunction Text

The function ID.

updateFunction_dataSourceName :: Lens' UpdateFunction Text

The Function DataSource name.

updateFunction_functionVersion :: Lens' UpdateFunction Text

The version of the request mapping template. Currently the supported value is 2018-05-29.

Destructuring the Response

data UpdateFunctionResponse

See: newUpdateFunctionResponse smart constructor.

Constructors

UpdateFunctionResponse'

Fields

functionConfiguration :: Maybe FunctionConfiguration

The Function object.

httpStatus :: Int

The response's http status code.

newUpdateFunctionResponse

Arguments

:: Int
$sel:httpStatus:UpdateFunctionResponse'

-> UpdateFunctionResponse

Create a value of UpdateFunctionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:functionConfiguration:UpdateFunctionResponse', updateFunctionResponse_functionConfiguration - The Function object.

$sel:httpStatus:UpdateFunctionResponse', updateFunctionResponse_httpStatus - The response's http status code.

Response Lenses

updateFunctionResponse_functionConfiguration :: Lens' UpdateFunctionResponse (Maybe FunctionConfiguration)

The Function object.

updateFunctionResponse_httpStatus :: Lens' UpdateFunctionResponse Int

The response's http status code.