Network.AWS.AppSync.UpdateGraphqlApi

Description

Updates a GraphqlApi object.

Creating a Request

data UpdateGraphqlApi

See: newUpdateGraphqlApi smart constructor.

Constructors

UpdateGraphqlApi'

Fields

openIDConnectConfig :: Maybe OpenIDConnectConfig

The OpenID Connect configuration for the GraphqlApi object.

userPoolConfig :: Maybe UserPoolConfig

The new Amazon Cognito user pool configuration for the GraphqlApi object.

xrayEnabled :: Maybe Bool

A flag indicating whether to enable X-Ray tracing for the GraphqlApi.

lambdaAuthorizerConfig :: Maybe LambdaAuthorizerConfig

Configuration for Amazon Web Services Lambda function authorization.

logConfig :: Maybe LogConfig

The Amazon CloudWatch Logs configuration for the GraphqlApi object.

additionalAuthenticationProviders :: Maybe [AdditionalAuthenticationProvider]

A list of additional authentication providers for the GraphqlApi API.

authenticationType :: Maybe AuthenticationType

The new authentication type for the GraphqlApi object.

apiId :: Text

The API ID.

name :: Text

The new name for the GraphqlApi object.

newUpdateGraphqlApi

Arguments

:: Text
$sel:apiId:UpdateGraphqlApi'

-> Text
$sel:name:UpdateGraphqlApi'

-> UpdateGraphqlApi

Create a value of UpdateGraphqlApi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:openIDConnectConfig:UpdateGraphqlApi', updateGraphqlApi_openIDConnectConfig - The OpenID Connect configuration for the GraphqlApi object.

$sel:userPoolConfig:UpdateGraphqlApi', updateGraphqlApi_userPoolConfig - The new Amazon Cognito user pool configuration for the GraphqlApi object.

$sel:xrayEnabled:UpdateGraphqlApi', updateGraphqlApi_xrayEnabled - A flag indicating whether to enable X-Ray tracing for the GraphqlApi.

$sel:lambdaAuthorizerConfig:UpdateGraphqlApi', updateGraphqlApi_lambdaAuthorizerConfig - Configuration for Amazon Web Services Lambda function authorization.

$sel:logConfig:UpdateGraphqlApi', updateGraphqlApi_logConfig - The Amazon CloudWatch Logs configuration for the GraphqlApi object.

$sel:additionalAuthenticationProviders:UpdateGraphqlApi', updateGraphqlApi_additionalAuthenticationProviders - A list of additional authentication providers for the GraphqlApi API.

$sel:authenticationType:UpdateGraphqlApi', updateGraphqlApi_authenticationType - The new authentication type for the GraphqlApi object.

$sel:apiId:UpdateGraphqlApi', updateGraphqlApi_apiId - The API ID.

$sel:name:UpdateGraphqlApi', updateGraphqlApi_name - The new name for the GraphqlApi object.

Request Lenses

updateGraphqlApi_openIDConnectConfig :: Lens' UpdateGraphqlApi (Maybe OpenIDConnectConfig)

The OpenID Connect configuration for the GraphqlApi object.

updateGraphqlApi_userPoolConfig :: Lens' UpdateGraphqlApi (Maybe UserPoolConfig)

The new Amazon Cognito user pool configuration for the GraphqlApi object.

updateGraphqlApi_xrayEnabled :: Lens' UpdateGraphqlApi (Maybe Bool)

A flag indicating whether to enable X-Ray tracing for the GraphqlApi.

updateGraphqlApi_lambdaAuthorizerConfig :: Lens' UpdateGraphqlApi (Maybe LambdaAuthorizerConfig)

Configuration for Amazon Web Services Lambda function authorization.

updateGraphqlApi_logConfig :: Lens' UpdateGraphqlApi (Maybe LogConfig)

The Amazon CloudWatch Logs configuration for the GraphqlApi object.

updateGraphqlApi_additionalAuthenticationProviders :: Lens' UpdateGraphqlApi (Maybe [AdditionalAuthenticationProvider])

A list of additional authentication providers for the GraphqlApi API.

updateGraphqlApi_authenticationType :: Lens' UpdateGraphqlApi (Maybe AuthenticationType)

The new authentication type for the GraphqlApi object.

updateGraphqlApi_apiId :: Lens' UpdateGraphqlApi Text

The API ID.

updateGraphqlApi_name :: Lens' UpdateGraphqlApi Text

The new name for the GraphqlApi object.

Destructuring the Response

data UpdateGraphqlApiResponse

See: newUpdateGraphqlApiResponse smart constructor.

Constructors

UpdateGraphqlApiResponse'

Fields

graphqlApi :: Maybe GraphqlApi

The updated GraphqlApi object.

httpStatus :: Int

The response's http status code.

newUpdateGraphqlApiResponse

Arguments

:: Int
$sel:httpStatus:UpdateGraphqlApiResponse'

-> UpdateGraphqlApiResponse

Create a value of UpdateGraphqlApiResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:graphqlApi:UpdateGraphqlApiResponse', updateGraphqlApiResponse_graphqlApi - The updated GraphqlApi object.

$sel:httpStatus:UpdateGraphqlApiResponse', updateGraphqlApiResponse_httpStatus - The response's http status code.

Response Lenses

updateGraphqlApiResponse_graphqlApi :: Lens' UpdateGraphqlApiResponse (Maybe GraphqlApi)

The updated GraphqlApi object.

updateGraphqlApiResponse_httpStatus :: Lens' UpdateGraphqlApiResponse Int

The response's http status code.