Network.AWS.AppSync.UpdateResolver

Description

Updates a Resolver object.

Creating a Request

data UpdateResolver

See: newUpdateResolver smart constructor.

Constructors

UpdateResolver'

Fields

dataSourceName :: Maybe Text

The new data source name.

requestMappingTemplate :: Maybe Text

The new request mapping template.

A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL).

VTL request mapping templates are optional when using a Lambda data source. For all other data sources, VTL request and response mapping templates are required.

kind :: Maybe ResolverKind

The resolver type.

UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source.

PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

cachingConfig :: Maybe CachingConfig

The caching configuration for the resolver.

responseMappingTemplate :: Maybe Text

The new response mapping template.

syncConfig :: Maybe SyncConfig

The SyncConfig for a resolver attached to a versioned datasource.

pipelineConfig :: Maybe PipelineConfig

The PipelineConfig.

apiId :: Text

The API ID.

typeName :: Text

The new type name.

fieldName :: Text

The new field name.

newUpdateResolver

Arguments

:: Text
$sel:apiId:UpdateResolver'

-> Text
$sel:typeName:UpdateResolver'

-> Text
$sel:fieldName:UpdateResolver'

-> UpdateResolver

Create a value of UpdateResolver with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSourceName:UpdateResolver', updateResolver_dataSourceName - The new data source name.

$sel:requestMappingTemplate:UpdateResolver', updateResolver_requestMappingTemplate - The new request mapping template.

A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL).

VTL request mapping templates are optional when using a Lambda data source. For all other data sources, VTL request and response mapping templates are required.

$sel:kind:UpdateResolver', updateResolver_kind - The resolver type.

UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source.

PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

$sel:cachingConfig:UpdateResolver', updateResolver_cachingConfig - The caching configuration for the resolver.

$sel:responseMappingTemplate:UpdateResolver', updateResolver_responseMappingTemplate - The new response mapping template.

$sel:syncConfig:UpdateResolver', updateResolver_syncConfig - The SyncConfig for a resolver attached to a versioned datasource.

$sel:pipelineConfig:UpdateResolver', updateResolver_pipelineConfig - The PipelineConfig.

$sel:apiId:UpdateResolver', updateResolver_apiId - The API ID.

$sel:typeName:UpdateResolver', updateResolver_typeName - The new type name.

$sel:fieldName:UpdateResolver', updateResolver_fieldName - The new field name.

Request Lenses

updateResolver_dataSourceName :: Lens' UpdateResolver (Maybe Text)

The new data source name.

updateResolver_requestMappingTemplate :: Lens' UpdateResolver (Maybe Text)

The new request mapping template.

A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL).

VTL request mapping templates are optional when using a Lambda data source. For all other data sources, VTL request and response mapping templates are required.

updateResolver_kind :: Lens' UpdateResolver (Maybe ResolverKind)

The resolver type.

UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source.

PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

updateResolver_cachingConfig :: Lens' UpdateResolver (Maybe CachingConfig)

The caching configuration for the resolver.

updateResolver_responseMappingTemplate :: Lens' UpdateResolver (Maybe Text)

The new response mapping template.

updateResolver_syncConfig :: Lens' UpdateResolver (Maybe SyncConfig)

The SyncConfig for a resolver attached to a versioned datasource.

updateResolver_pipelineConfig :: Lens' UpdateResolver (Maybe PipelineConfig)

The PipelineConfig.

updateResolver_apiId :: Lens' UpdateResolver Text

The API ID.

updateResolver_typeName :: Lens' UpdateResolver Text

The new type name.

updateResolver_fieldName :: Lens' UpdateResolver Text

The new field name.

Destructuring the Response

data UpdateResolverResponse

See: newUpdateResolverResponse smart constructor.

Constructors

UpdateResolverResponse'

Fields

resolver :: Maybe Resolver

The updated Resolver object.

httpStatus :: Int

The response's http status code.

newUpdateResolverResponse

Arguments

:: Int
$sel:httpStatus:UpdateResolverResponse'

-> UpdateResolverResponse

Create a value of UpdateResolverResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolver:UpdateResolverResponse', updateResolverResponse_resolver - The updated Resolver object.

$sel:httpStatus:UpdateResolverResponse', updateResolverResponse_httpStatus - The response's http status code.

Response Lenses

updateResolverResponse_resolver :: Lens' UpdateResolverResponse (Maybe Resolver)

The updated Resolver object.

updateResolverResponse_httpStatus :: Lens' UpdateResolverResponse Int

The response's http status code.