Network.AWS.ApplicationAutoScaling.Types.ScalingActivity

Description

Documentation

data ScalingActivity

Represents a scaling activity.

See: newScalingActivity smart constructor.

Constructors

ScalingActivity'

Fields

statusMessage :: Maybe Text

A simple message about the current status of the scaling activity.

endTime :: Maybe POSIX

The Unix timestamp for when the scaling activity ended.

details :: Maybe Text

The details about the scaling activity.

activityId :: Text

The unique identifier of the scaling activity.

serviceNamespace :: ServiceNamespace

The namespace of the Amazon Web Services service that provides the resource, or a custom-resource.

resourceId :: Text

The identifier of the resource associated with the scaling activity. This string consists of the resource type and unique identifier.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the table name. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the index name. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

Amazon SageMaker endpoint variant - The resource type is variant and the unique identifier is the resource ID. Example: endpoint/my-end-point/variant/KMeansClustering.

Custom resources are not supported with a resource type. This parameter must specify the OutputValue from the CloudFormation template stack used to access the resources. The unique identifier is defined by the service provider. More information is available in our GitHub repository.

Amazon Comprehend document classification endpoint - The resource type and unique identifier are specified using the endpoint ARN. Example: arn:aws:comprehend:us-west-2:123456789012:document-classifier-endpoint/EXAMPLE.

Amazon Comprehend entity recognizer endpoint - The resource type and unique identifier are specified using the endpoint ARN. Example: arn:aws:comprehend:us-west-2:123456789012:entity-recognizer-endpoint/EXAMPLE.

Lambda provisioned concurrency - The resource type is function and the unique identifier is the function name with a function version or alias name suffix that is not $LATEST. Example: function:my-function:prod or function:my-function:1.

Amazon Keyspaces table - The resource type is table and the unique identifier is the table name. Example: keyspace/mykeyspace/table/mytable.

Amazon MSK cluster - The resource type and unique identifier are specified using the cluster ARN. Example: arn:aws:kafka:us-east-1:123456789012:cluster/demo-cluster-1/6357e0b2-0e6a-4b86-a0b4-70df934c2e31-5.

Amazon ElastiCache replication group - The resource type is replication-group and the unique identifier is the replication group name. Example: replication-group/mycluster.

scalableDimension :: ScalableDimension

The scalable dimension. This string consists of the service namespace, resource type, and scaling property.

ecs:service:DesiredCount - The desired task count of an ECS service.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet request.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

sagemaker:variant:DesiredInstanceCount - The number of EC2 instances for an Amazon SageMaker model endpoint variant.

custom-resource:ResourceType:Property - The scalable dimension for a custom resource provided by your own application or service.

comprehend:document-classifier-endpoint:DesiredInferenceUnits - The number of inference units for an Amazon Comprehend document classification endpoint.

comprehend:entity-recognizer-endpoint:DesiredInferenceUnits - The number of inference units for an Amazon Comprehend entity recognizer endpoint.

lambda:function:ProvisionedConcurrency - The provisioned concurrency for a Lambda function.

cassandra:table:ReadCapacityUnits - The provisioned read capacity for an Amazon Keyspaces table.

cassandra:table:WriteCapacityUnits - The provisioned write capacity for an Amazon Keyspaces table.

kafka:broker-storage:VolumeSize - The provisioned volume size (in GiB) for brokers in an Amazon MSK cluster.

elasticache:replication-group:NodeGroups - The number of node groups for an Amazon ElastiCache replication group.

elasticache:replication-group:Replicas - The number of replicas per node group for an Amazon ElastiCache replication group.

description :: Text

A simple description of what action the scaling activity intends to accomplish.

cause :: Text

A simple description of what caused the scaling activity to happen.

startTime :: POSIX

The Unix timestamp for when the scaling activity began.

statusCode :: ScalingActivityStatusCode

Indicates the status of the scaling activity.

newScalingActivity

Arguments

:: Text
$sel:activityId:ScalingActivity'

-> ServiceNamespace
$sel:serviceNamespace:ScalingActivity'

-> Text
$sel:resourceId:ScalingActivity'

-> ScalableDimension
$sel:scalableDimension:ScalingActivity'

-> Text
$sel:description:ScalingActivity'

-> Text
$sel:cause:ScalingActivity'

-> UTCTime
$sel:startTime:ScalingActivity'

-> ScalingActivityStatusCode
$sel:statusCode:ScalingActivity'

-> ScalingActivity

Create a value of ScalingActivity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:ScalingActivity', scalingActivity_statusMessage - A simple message about the current status of the scaling activity.

$sel:endTime:ScalingActivity', scalingActivity_endTime - The Unix timestamp for when the scaling activity ended.

$sel:details:ScalingActivity', scalingActivity_details - The details about the scaling activity.

$sel:activityId:ScalingActivity', scalingActivity_activityId - The unique identifier of the scaling activity.

$sel:serviceNamespace:ScalingActivity', scalingActivity_serviceNamespace - The namespace of the Amazon Web Services service that provides the resource, or a custom-resource.

$sel:resourceId:ScalingActivity', scalingActivity_resourceId - The identifier of the resource associated with the scaling activity. This string consists of the resource type and unique identifier.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the table name. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the index name. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

Amazon SageMaker endpoint variant - The resource type is variant and the unique identifier is the resource ID. Example: endpoint/my-end-point/variant/KMeansClustering.

Custom resources are not supported with a resource type. This parameter must specify the OutputValue from the CloudFormation template stack used to access the resources. The unique identifier is defined by the service provider. More information is available in our GitHub repository.

Amazon Comprehend document classification endpoint - The resource type and unique identifier are specified using the endpoint ARN. Example: arn:aws:comprehend:us-west-2:123456789012:document-classifier-endpoint/EXAMPLE.

Amazon Comprehend entity recognizer endpoint - The resource type and unique identifier are specified using the endpoint ARN. Example: arn:aws:comprehend:us-west-2:123456789012:entity-recognizer-endpoint/EXAMPLE.

Lambda provisioned concurrency - The resource type is function and the unique identifier is the function name with a function version or alias name suffix that is not $LATEST. Example: function:my-function:prod or function:my-function:1.

Amazon Keyspaces table - The resource type is table and the unique identifier is the table name. Example: keyspace/mykeyspace/table/mytable.

Amazon MSK cluster - The resource type and unique identifier are specified using the cluster ARN. Example: arn:aws:kafka:us-east-1:123456789012:cluster/demo-cluster-1/6357e0b2-0e6a-4b86-a0b4-70df934c2e31-5.

Amazon ElastiCache replication group - The resource type is replication-group and the unique identifier is the replication group name. Example: replication-group/mycluster.

$sel:scalableDimension:ScalingActivity', scalingActivity_scalableDimension - The scalable dimension. This string consists of the service namespace, resource type, and scaling property.

ecs:service:DesiredCount - The desired task count of an ECS service.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet request.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

sagemaker:variant:DesiredInstanceCount - The number of EC2 instances for an Amazon SageMaker model endpoint variant.

custom-resource:ResourceType:Property - The scalable dimension for a custom resource provided by your own application or service.

comprehend:document-classifier-endpoint:DesiredInferenceUnits - The number of inference units for an Amazon Comprehend document classification endpoint.

comprehend:entity-recognizer-endpoint:DesiredInferenceUnits - The number of inference units for an Amazon Comprehend entity recognizer endpoint.

lambda:function:ProvisionedConcurrency - The provisioned concurrency for a Lambda function.

cassandra:table:ReadCapacityUnits - The provisioned read capacity for an Amazon Keyspaces table.

cassandra:table:WriteCapacityUnits - The provisioned write capacity for an Amazon Keyspaces table.

kafka:broker-storage:VolumeSize - The provisioned volume size (in GiB) for brokers in an Amazon MSK cluster.

elasticache:replication-group:NodeGroups - The number of node groups for an Amazon ElastiCache replication group.

elasticache:replication-group:Replicas - The number of replicas per node group for an Amazon ElastiCache replication group.

$sel:description:ScalingActivity', scalingActivity_description - A simple description of what action the scaling activity intends to accomplish.

$sel:cause:ScalingActivity', scalingActivity_cause - A simple description of what caused the scaling activity to happen.

$sel:startTime:ScalingActivity', scalingActivity_startTime - The Unix timestamp for when the scaling activity began.

$sel:statusCode:ScalingActivity', scalingActivity_statusCode - Indicates the status of the scaling activity.

scalingActivity_statusMessage :: Lens' ScalingActivity (Maybe Text)

A simple message about the current status of the scaling activity.

scalingActivity_endTime :: Lens' ScalingActivity (Maybe UTCTime)

The Unix timestamp for when the scaling activity ended.

scalingActivity_details :: Lens' ScalingActivity (Maybe Text)

The details about the scaling activity.

scalingActivity_activityId :: Lens' ScalingActivity Text

The unique identifier of the scaling activity.

scalingActivity_serviceNamespace :: Lens' ScalingActivity ServiceNamespace

The namespace of the Amazon Web Services service that provides the resource, or a custom-resource.

scalingActivity_resourceId :: Lens' ScalingActivity Text

The identifier of the resource associated with the scaling activity. This string consists of the resource type and unique identifier.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the table name. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the index name. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

Amazon SageMaker endpoint variant - The resource type is variant and the unique identifier is the resource ID. Example: endpoint/my-end-point/variant/KMeansClustering.

Custom resources are not supported with a resource type. This parameter must specify the OutputValue from the CloudFormation template stack used to access the resources. The unique identifier is defined by the service provider. More information is available in our GitHub repository.

Amazon Comprehend document classification endpoint - The resource type and unique identifier are specified using the endpoint ARN. Example: arn:aws:comprehend:us-west-2:123456789012:document-classifier-endpoint/EXAMPLE.

Amazon Comprehend entity recognizer endpoint - The resource type and unique identifier are specified using the endpoint ARN. Example: arn:aws:comprehend:us-west-2:123456789012:entity-recognizer-endpoint/EXAMPLE.

Lambda provisioned concurrency - The resource type is function and the unique identifier is the function name with a function version or alias name suffix that is not $LATEST. Example: function:my-function:prod or function:my-function:1.

Amazon Keyspaces table - The resource type is table and the unique identifier is the table name. Example: keyspace/mykeyspace/table/mytable.

Amazon MSK cluster - The resource type and unique identifier are specified using the cluster ARN. Example: arn:aws:kafka:us-east-1:123456789012:cluster/demo-cluster-1/6357e0b2-0e6a-4b86-a0b4-70df934c2e31-5.

Amazon ElastiCache replication group - The resource type is replication-group and the unique identifier is the replication group name. Example: replication-group/mycluster.

scalingActivity_scalableDimension :: Lens' ScalingActivity ScalableDimension

The scalable dimension. This string consists of the service namespace, resource type, and scaling property.

ecs:service:DesiredCount - The desired task count of an ECS service.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet request.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

sagemaker:variant:DesiredInstanceCount - The number of EC2 instances for an Amazon SageMaker model endpoint variant.

custom-resource:ResourceType:Property - The scalable dimension for a custom resource provided by your own application or service.

comprehend:document-classifier-endpoint:DesiredInferenceUnits - The number of inference units for an Amazon Comprehend document classification endpoint.

comprehend:entity-recognizer-endpoint:DesiredInferenceUnits - The number of inference units for an Amazon Comprehend entity recognizer endpoint.

lambda:function:ProvisionedConcurrency - The provisioned concurrency for a Lambda function.

cassandra:table:ReadCapacityUnits - The provisioned read capacity for an Amazon Keyspaces table.

cassandra:table:WriteCapacityUnits - The provisioned write capacity for an Amazon Keyspaces table.

kafka:broker-storage:VolumeSize - The provisioned volume size (in GiB) for brokers in an Amazon MSK cluster.

elasticache:replication-group:NodeGroups - The number of node groups for an Amazon ElastiCache replication group.

elasticache:replication-group:Replicas - The number of replicas per node group for an Amazon ElastiCache replication group.

scalingActivity_description :: Lens' ScalingActivity Text

A simple description of what action the scaling activity intends to accomplish.

scalingActivity_cause :: Lens' ScalingActivity Text

A simple description of what caused the scaling activity to happen.

scalingActivity_startTime :: Lens' ScalingActivity UTCTime

The Unix timestamp for when the scaling activity began.

scalingActivity_statusCode :: Lens' ScalingActivity ScalingActivityStatusCode

Indicates the status of the scaling activity.