Network.AWS.ApplicationAutoScaling.Types.ScalingPolicy

Description

Documentation

data ScalingPolicy

Represents a scaling policy to use with Application Auto Scaling.

For more information about configuring scaling policies for a specific service, see Getting started with Application Auto Scaling in the Application Auto Scaling User Guide.

See: newScalingPolicy smart constructor.

Constructors

ScalingPolicy'

Fields

targetTrackingScalingPolicyConfiguration :: Maybe TargetTrackingScalingPolicyConfiguration

A target tracking scaling policy.

stepScalingPolicyConfiguration :: Maybe StepScalingPolicyConfiguration

A step scaling policy.

alarms :: Maybe [Alarm]

The CloudWatch alarms associated with the scaling policy.

policyARN :: Text

The Amazon Resource Name (ARN) of the scaling policy.

policyName :: Text

The name of the scaling policy.

serviceNamespace :: ServiceNamespace

The namespace of the Amazon Web Services service that provides the resource, or a custom-resource.

resourceId :: Text

The identifier of the resource associated with the scaling policy. This string consists of the resource type and unique identifier.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the table name. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the index name. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

SageMaker endpoint variant - The resource type is variant and the unique identifier is the resource ID. Example: endpoint/my-end-point/variant/KMeansClustering.

Custom resources are not supported with a resource type. This parameter must specify the OutputValue from the CloudFormation template stack used to access the resources. The unique identifier is defined by the service provider. More information is available in our GitHub repository.

Amazon Comprehend document classification endpoint - The resource type and unique identifier are specified using the endpoint ARN. Example: arn:aws:comprehend:us-west-2:123456789012:document-classifier-endpoint/EXAMPLE.

Amazon Comprehend entity recognizer endpoint - The resource type and unique identifier are specified using the endpoint ARN. Example: arn:aws:comprehend:us-west-2:123456789012:entity-recognizer-endpoint/EXAMPLE.

Lambda provisioned concurrency - The resource type is function and the unique identifier is the function name with a function version or alias name suffix that is not $LATEST. Example: function:my-function:prod or function:my-function:1.

Amazon Keyspaces table - The resource type is table and the unique identifier is the table name. Example: keyspace/mykeyspace/table/mytable.

Amazon MSK cluster - The resource type and unique identifier are specified using the cluster ARN. Example: arn:aws:kafka:us-east-1:123456789012:cluster/demo-cluster-1/6357e0b2-0e6a-4b86-a0b4-70df934c2e31-5.

Amazon ElastiCache replication group - The resource type is replication-group and the unique identifier is the replication group name. Example: replication-group/mycluster.

Neptune cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:mycluster.

scalableDimension :: ScalableDimension

The scalable dimension. This string consists of the service namespace, resource type, and scaling property.

ecs:service:DesiredCount - The desired task count of an ECS service.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

sagemaker:variant:DesiredInstanceCount - The number of EC2 instances for an SageMaker model endpoint variant.

custom-resource:ResourceType:Property - The scalable dimension for a custom resource provided by your own application or service.

comprehend:document-classifier-endpoint:DesiredInferenceUnits - The number of inference units for an Amazon Comprehend document classification endpoint.

comprehend:entity-recognizer-endpoint:DesiredInferenceUnits - The number of inference units for an Amazon Comprehend entity recognizer endpoint.

lambda:function:ProvisionedConcurrency - The provisioned concurrency for a Lambda function.

cassandra:table:ReadCapacityUnits - The provisioned read capacity for an Amazon Keyspaces table.

cassandra:table:WriteCapacityUnits - The provisioned write capacity for an Amazon Keyspaces table.

kafka:broker-storage:VolumeSize - The provisioned volume size (in GiB) for brokers in an Amazon MSK cluster.

elasticache:replication-group:NodeGroups - The number of node groups for an Amazon ElastiCache replication group.

elasticache:replication-group:Replicas - The number of replicas per node group for an Amazon ElastiCache replication group.

neptune:cluster:ReadReplicaCount - The count of read replicas in an Amazon Neptune DB cluster.

policyType :: PolicyType

The scaling policy type.

creationTime :: POSIX

The Unix timestamp for when the scaling policy was created.

newScalingPolicy

Arguments

:: Text
$sel:policyARN:ScalingPolicy'

-> Text
$sel:policyName:ScalingPolicy'

-> ServiceNamespace
$sel:serviceNamespace:ScalingPolicy'

-> Text
$sel:resourceId:ScalingPolicy'

-> ScalableDimension
$sel:scalableDimension:ScalingPolicy'

-> PolicyType
$sel:policyType:ScalingPolicy'

-> UTCTime
$sel:creationTime:ScalingPolicy'

-> ScalingPolicy

Create a value of ScalingPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetTrackingScalingPolicyConfiguration:ScalingPolicy', scalingPolicy_targetTrackingScalingPolicyConfiguration - A target tracking scaling policy.

$sel:stepScalingPolicyConfiguration:ScalingPolicy', scalingPolicy_stepScalingPolicyConfiguration - A step scaling policy.

$sel:alarms:ScalingPolicy', scalingPolicy_alarms - The CloudWatch alarms associated with the scaling policy.

$sel:policyARN:ScalingPolicy', scalingPolicy_policyARN - The Amazon Resource Name (ARN) of the scaling policy.

$sel:policyName:ScalingPolicy', scalingPolicy_policyName - The name of the scaling policy.

$sel:serviceNamespace:ScalingPolicy', scalingPolicy_serviceNamespace - The namespace of the Amazon Web Services service that provides the resource, or a custom-resource.

$sel:resourceId:ScalingPolicy', scalingPolicy_resourceId - The identifier of the resource associated with the scaling policy. This string consists of the resource type and unique identifier.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the table name. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the index name. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

SageMaker endpoint variant - The resource type is variant and the unique identifier is the resource ID. Example: endpoint/my-end-point/variant/KMeansClustering.

Custom resources are not supported with a resource type. This parameter must specify the OutputValue from the CloudFormation template stack used to access the resources. The unique identifier is defined by the service provider. More information is available in our GitHub repository.

Amazon Comprehend document classification endpoint - The resource type and unique identifier are specified using the endpoint ARN. Example: arn:aws:comprehend:us-west-2:123456789012:document-classifier-endpoint/EXAMPLE.

Amazon Comprehend entity recognizer endpoint - The resource type and unique identifier are specified using the endpoint ARN. Example: arn:aws:comprehend:us-west-2:123456789012:entity-recognizer-endpoint/EXAMPLE.

Lambda provisioned concurrency - The resource type is function and the unique identifier is the function name with a function version or alias name suffix that is not $LATEST. Example: function:my-function:prod or function:my-function:1.

Amazon Keyspaces table - The resource type is table and the unique identifier is the table name. Example: keyspace/mykeyspace/table/mytable.

Amazon MSK cluster - The resource type and unique identifier are specified using the cluster ARN. Example: arn:aws:kafka:us-east-1:123456789012:cluster/demo-cluster-1/6357e0b2-0e6a-4b86-a0b4-70df934c2e31-5.

Amazon ElastiCache replication group - The resource type is replication-group and the unique identifier is the replication group name. Example: replication-group/mycluster.

Neptune cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:mycluster.

$sel:scalableDimension:ScalingPolicy', scalingPolicy_scalableDimension - The scalable dimension. This string consists of the service namespace, resource type, and scaling property.

ecs:service:DesiredCount - The desired task count of an ECS service.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

sagemaker:variant:DesiredInstanceCount - The number of EC2 instances for an SageMaker model endpoint variant.

custom-resource:ResourceType:Property - The scalable dimension for a custom resource provided by your own application or service.

comprehend:document-classifier-endpoint:DesiredInferenceUnits - The number of inference units for an Amazon Comprehend document classification endpoint.

comprehend:entity-recognizer-endpoint:DesiredInferenceUnits - The number of inference units for an Amazon Comprehend entity recognizer endpoint.

lambda:function:ProvisionedConcurrency - The provisioned concurrency for a Lambda function.

cassandra:table:ReadCapacityUnits - The provisioned read capacity for an Amazon Keyspaces table.

cassandra:table:WriteCapacityUnits - The provisioned write capacity for an Amazon Keyspaces table.

kafka:broker-storage:VolumeSize - The provisioned volume size (in GiB) for brokers in an Amazon MSK cluster.

elasticache:replication-group:NodeGroups - The number of node groups for an Amazon ElastiCache replication group.

elasticache:replication-group:Replicas - The number of replicas per node group for an Amazon ElastiCache replication group.

neptune:cluster:ReadReplicaCount - The count of read replicas in an Amazon Neptune DB cluster.

$sel:policyType:ScalingPolicy', scalingPolicy_policyType - The scaling policy type.

$sel:creationTime:ScalingPolicy', scalingPolicy_creationTime - The Unix timestamp for when the scaling policy was created.

scalingPolicy_targetTrackingScalingPolicyConfiguration :: Lens' ScalingPolicy (Maybe TargetTrackingScalingPolicyConfiguration)

A target tracking scaling policy.

scalingPolicy_stepScalingPolicyConfiguration :: Lens' ScalingPolicy (Maybe StepScalingPolicyConfiguration)

A step scaling policy.

scalingPolicy_alarms :: Lens' ScalingPolicy (Maybe [Alarm])

The CloudWatch alarms associated with the scaling policy.

scalingPolicy_policyARN :: Lens' ScalingPolicy Text

The Amazon Resource Name (ARN) of the scaling policy.

scalingPolicy_policyName :: Lens' ScalingPolicy Text

The name of the scaling policy.

scalingPolicy_serviceNamespace :: Lens' ScalingPolicy ServiceNamespace

The namespace of the Amazon Web Services service that provides the resource, or a custom-resource.

scalingPolicy_resourceId :: Lens' ScalingPolicy Text

The identifier of the resource associated with the scaling policy. This string consists of the resource type and unique identifier.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the table name. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the index name. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

SageMaker endpoint variant - The resource type is variant and the unique identifier is the resource ID. Example: endpoint/my-end-point/variant/KMeansClustering.

Custom resources are not supported with a resource type. This parameter must specify the OutputValue from the CloudFormation template stack used to access the resources. The unique identifier is defined by the service provider. More information is available in our GitHub repository.

Amazon Comprehend document classification endpoint - The resource type and unique identifier are specified using the endpoint ARN. Example: arn:aws:comprehend:us-west-2:123456789012:document-classifier-endpoint/EXAMPLE.

Amazon Comprehend entity recognizer endpoint - The resource type and unique identifier are specified using the endpoint ARN. Example: arn:aws:comprehend:us-west-2:123456789012:entity-recognizer-endpoint/EXAMPLE.

Lambda provisioned concurrency - The resource type is function and the unique identifier is the function name with a function version or alias name suffix that is not $LATEST. Example: function:my-function:prod or function:my-function:1.

Amazon Keyspaces table - The resource type is table and the unique identifier is the table name. Example: keyspace/mykeyspace/table/mytable.

Amazon MSK cluster - The resource type and unique identifier are specified using the cluster ARN. Example: arn:aws:kafka:us-east-1:123456789012:cluster/demo-cluster-1/6357e0b2-0e6a-4b86-a0b4-70df934c2e31-5.

Amazon ElastiCache replication group - The resource type is replication-group and the unique identifier is the replication group name. Example: replication-group/mycluster.

Neptune cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:mycluster.

scalingPolicy_scalableDimension :: Lens' ScalingPolicy ScalableDimension

The scalable dimension. This string consists of the service namespace, resource type, and scaling property.

ecs:service:DesiredCount - The desired task count of an ECS service.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

sagemaker:variant:DesiredInstanceCount - The number of EC2 instances for an SageMaker model endpoint variant.

custom-resource:ResourceType:Property - The scalable dimension for a custom resource provided by your own application or service.

comprehend:document-classifier-endpoint:DesiredInferenceUnits - The number of inference units for an Amazon Comprehend document classification endpoint.

comprehend:entity-recognizer-endpoint:DesiredInferenceUnits - The number of inference units for an Amazon Comprehend entity recognizer endpoint.

lambda:function:ProvisionedConcurrency - The provisioned concurrency for a Lambda function.

cassandra:table:ReadCapacityUnits - The provisioned read capacity for an Amazon Keyspaces table.

cassandra:table:WriteCapacityUnits - The provisioned write capacity for an Amazon Keyspaces table.

kafka:broker-storage:VolumeSize - The provisioned volume size (in GiB) for brokers in an Amazon MSK cluster.

elasticache:replication-group:NodeGroups - The number of node groups for an Amazon ElastiCache replication group.

elasticache:replication-group:Replicas - The number of replicas per node group for an Amazon ElastiCache replication group.

neptune:cluster:ReadReplicaCount - The count of read replicas in an Amazon Neptune DB cluster.

scalingPolicy_policyType :: Lens' ScalingPolicy PolicyType

The scaling policy type.

scalingPolicy_creationTime :: Lens' ScalingPolicy UTCTime

The Unix timestamp for when the scaling policy was created.