Network.AWS.ApplicationAutoScaling.Types.TargetTrackingScalingPolicyConfiguration

Description

Documentation

data TargetTrackingScalingPolicyConfiguration

Represents a target tracking scaling policy configuration to use with Application Auto Scaling.

See: newTargetTrackingScalingPolicyConfiguration smart constructor.

Constructors

TargetTrackingScalingPolicyConfiguration'

Fields

disableScaleIn :: Maybe Bool

Indicates whether scale in by the target tracking scaling policy is disabled. If the value is true, scale in is disabled and the target tracking scaling policy won't remove capacity from the scalable target. Otherwise, scale in is enabled and the target tracking scaling policy can remove capacity from the scalable target. The default value is false.

predefinedMetricSpecification :: Maybe PredefinedMetricSpecification

A predefined metric. You can specify either a predefined metric or a customized metric.

scaleOutCooldown :: Maybe Int

The amount of time, in seconds, to wait for a previous scale-out activity to take effect.

With the scale-out cooldown period, the intention is to continuously (but not excessively) scale out. After Application Auto Scaling successfully scales out using a target tracking scaling policy, it starts to calculate the cooldown time. The scaling policy won't increase the desired capacity again unless either a larger scale out is triggered or the cooldown period ends. While the cooldown period is in effect, the capacity added by the initiating scale-out activity is calculated as part of the desired capacity for the next scale-out activity.

Application Auto Scaling provides a default value of 600 for Amazon ElastiCache replication groups and a default value of 300 for the following scalable targets:

ECS services

Spot Fleet requests

EMR clusters

AppStream 2.0 fleets

Aurora DB clusters

Amazon SageMaker endpoint variants

Custom resources

For all other scalable targets, the default value is 0:

DynamoDB tables

DynamoDB global secondary indexes

Amazon Comprehend document classification and entity recognizer endpoints

Lambda provisioned concurrency

Amazon Keyspaces tables

Amazon MSK broker storage

customizedMetricSpecification :: Maybe CustomizedMetricSpecification

A customized metric. You can specify either a predefined metric or a customized metric.

scaleInCooldown :: Maybe Int

The amount of time, in seconds, after a scale-in activity completes before another scale-in activity can start.

With the scale-in cooldown period, the intention is to scale in conservatively to protect your application’s availability, so scale-in activities are blocked until the cooldown period has expired. However, if another alarm triggers a scale-out activity during the scale-in cooldown period, Application Auto Scaling scales out the target immediately. In this case, the scale-in cooldown period stops and doesn't complete.

Application Auto Scaling provides a default value of 600 for Amazon ElastiCache replication groups and a default value of 300 for the following scalable targets:

ECS services

Spot Fleet requests

EMR clusters

AppStream 2.0 fleets

Aurora DB clusters

Amazon SageMaker endpoint variants

Custom resources

For all other scalable targets, the default value is 0:

DynamoDB tables

DynamoDB global secondary indexes

Amazon Comprehend document classification and entity recognizer endpoints

Lambda provisioned concurrency

Amazon Keyspaces tables

Amazon MSK broker storage

targetValue :: Double

The target value for the metric. Although this property accepts numbers of type Double, it won't accept values that are either too small or too large. Values must be in the range of -2^360 to 2^360. The value must be a valid number based on the choice of metric. For example, if the metric is CPU utilization, then the target value is a percent value that represents how much of the CPU can be used before scaling out.

newTargetTrackingScalingPolicyConfiguration

Arguments

:: Double
$sel:targetValue:TargetTrackingScalingPolicyConfiguration'

-> TargetTrackingScalingPolicyConfiguration

Create a value of TargetTrackingScalingPolicyConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disableScaleIn:TargetTrackingScalingPolicyConfiguration', targetTrackingScalingPolicyConfiguration_disableScaleIn - Indicates whether scale in by the target tracking scaling policy is disabled. If the value is true, scale in is disabled and the target tracking scaling policy won't remove capacity from the scalable target. Otherwise, scale in is enabled and the target tracking scaling policy can remove capacity from the scalable target. The default value is false.

$sel:predefinedMetricSpecification:TargetTrackingScalingPolicyConfiguration', targetTrackingScalingPolicyConfiguration_predefinedMetricSpecification - A predefined metric. You can specify either a predefined metric or a customized metric.

$sel:scaleOutCooldown:TargetTrackingScalingPolicyConfiguration', targetTrackingScalingPolicyConfiguration_scaleOutCooldown - The amount of time, in seconds, to wait for a previous scale-out activity to take effect.

With the scale-out cooldown period, the intention is to continuously (but not excessively) scale out. After Application Auto Scaling successfully scales out using a target tracking scaling policy, it starts to calculate the cooldown time. The scaling policy won't increase the desired capacity again unless either a larger scale out is triggered or the cooldown period ends. While the cooldown period is in effect, the capacity added by the initiating scale-out activity is calculated as part of the desired capacity for the next scale-out activity.

Application Auto Scaling provides a default value of 600 for Amazon ElastiCache replication groups and a default value of 300 for the following scalable targets:

ECS services

Spot Fleet requests

EMR clusters

AppStream 2.0 fleets

Aurora DB clusters

Amazon SageMaker endpoint variants

Custom resources

For all other scalable targets, the default value is 0:

DynamoDB tables

DynamoDB global secondary indexes

Amazon Comprehend document classification and entity recognizer endpoints

Lambda provisioned concurrency

Amazon Keyspaces tables

Amazon MSK broker storage

$sel:customizedMetricSpecification:TargetTrackingScalingPolicyConfiguration', targetTrackingScalingPolicyConfiguration_customizedMetricSpecification - A customized metric. You can specify either a predefined metric or a customized metric.

$sel:scaleInCooldown:TargetTrackingScalingPolicyConfiguration', targetTrackingScalingPolicyConfiguration_scaleInCooldown - The amount of time, in seconds, after a scale-in activity completes before another scale-in activity can start.

With the scale-in cooldown period, the intention is to scale in conservatively to protect your application’s availability, so scale-in activities are blocked until the cooldown period has expired. However, if another alarm triggers a scale-out activity during the scale-in cooldown period, Application Auto Scaling scales out the target immediately. In this case, the scale-in cooldown period stops and doesn't complete.

Application Auto Scaling provides a default value of 600 for Amazon ElastiCache replication groups and a default value of 300 for the following scalable targets:

ECS services

Spot Fleet requests

EMR clusters

AppStream 2.0 fleets

Aurora DB clusters

Amazon SageMaker endpoint variants

Custom resources

For all other scalable targets, the default value is 0:

DynamoDB tables

DynamoDB global secondary indexes

Amazon Comprehend document classification and entity recognizer endpoints

Lambda provisioned concurrency

Amazon Keyspaces tables

Amazon MSK broker storage

$sel:targetValue:TargetTrackingScalingPolicyConfiguration', targetTrackingScalingPolicyConfiguration_targetValue - The target value for the metric. Although this property accepts numbers of type Double, it won't accept values that are either too small or too large. Values must be in the range of -2^360 to 2^360. The value must be a valid number based on the choice of metric. For example, if the metric is CPU utilization, then the target value is a percent value that represents how much of the CPU can be used before scaling out.

targetTrackingScalingPolicyConfiguration_disableScaleIn :: Lens' TargetTrackingScalingPolicyConfiguration (Maybe Bool)

Indicates whether scale in by the target tracking scaling policy is disabled. If the value is true, scale in is disabled and the target tracking scaling policy won't remove capacity from the scalable target. Otherwise, scale in is enabled and the target tracking scaling policy can remove capacity from the scalable target. The default value is false.

targetTrackingScalingPolicyConfiguration_predefinedMetricSpecification :: Lens' TargetTrackingScalingPolicyConfiguration (Maybe PredefinedMetricSpecification)

A predefined metric. You can specify either a predefined metric or a customized metric.

targetTrackingScalingPolicyConfiguration_scaleOutCooldown :: Lens' TargetTrackingScalingPolicyConfiguration (Maybe Int)

The amount of time, in seconds, to wait for a previous scale-out activity to take effect.

With the scale-out cooldown period, the intention is to continuously (but not excessively) scale out. After Application Auto Scaling successfully scales out using a target tracking scaling policy, it starts to calculate the cooldown time. The scaling policy won't increase the desired capacity again unless either a larger scale out is triggered or the cooldown period ends. While the cooldown period is in effect, the capacity added by the initiating scale-out activity is calculated as part of the desired capacity for the next scale-out activity.

Application Auto Scaling provides a default value of 600 for Amazon ElastiCache replication groups and a default value of 300 for the following scalable targets:

ECS services

Spot Fleet requests

EMR clusters

AppStream 2.0 fleets

Aurora DB clusters

Amazon SageMaker endpoint variants

Custom resources

For all other scalable targets, the default value is 0:

DynamoDB tables

DynamoDB global secondary indexes

Amazon Comprehend document classification and entity recognizer endpoints

Lambda provisioned concurrency

Amazon Keyspaces tables

Amazon MSK broker storage

targetTrackingScalingPolicyConfiguration_customizedMetricSpecification :: Lens' TargetTrackingScalingPolicyConfiguration (Maybe CustomizedMetricSpecification)

A customized metric. You can specify either a predefined metric or a customized metric.

targetTrackingScalingPolicyConfiguration_scaleInCooldown :: Lens' TargetTrackingScalingPolicyConfiguration (Maybe Int)

The amount of time, in seconds, after a scale-in activity completes before another scale-in activity can start.

With the scale-in cooldown period, the intention is to scale in conservatively to protect your application’s availability, so scale-in activities are blocked until the cooldown period has expired. However, if another alarm triggers a scale-out activity during the scale-in cooldown period, Application Auto Scaling scales out the target immediately. In this case, the scale-in cooldown period stops and doesn't complete.

Application Auto Scaling provides a default value of 600 for Amazon ElastiCache replication groups and a default value of 300 for the following scalable targets:

ECS services

Spot Fleet requests

EMR clusters

AppStream 2.0 fleets

Aurora DB clusters

Amazon SageMaker endpoint variants

Custom resources

For all other scalable targets, the default value is 0:

DynamoDB tables

DynamoDB global secondary indexes

Amazon Comprehend document classification and entity recognizer endpoints

Lambda provisioned concurrency

Amazon Keyspaces tables

Amazon MSK broker storage

targetTrackingScalingPolicyConfiguration_targetValue :: Lens' TargetTrackingScalingPolicyConfiguration Double

The target value for the metric. Although this property accepts numbers of type Double, it won't accept values that are either too small or too large. Values must be in the range of -2^360 to 2^360. The value must be a valid number based on the choice of metric. For example, if the metric is CPU utilization, then the target value is a percent value that represents how much of the CPU can be used before scaling out.