Network.AWS.ApplicationCostProfiler.GetReportDefinition

Description

Retrieves the definition of a report already configured in AWS Application Cost Profiler.

Creating a Request

data GetReportDefinition

See: newGetReportDefinition smart constructor.

Constructors

GetReportDefinition'

Fields

reportId :: Text

ID of the report to retrieve.

newGetReportDefinition

Arguments

:: Text
$sel:reportId:GetReportDefinition'

-> GetReportDefinition

Create a value of GetReportDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportId:GetReportDefinition', getReportDefinition_reportId - ID of the report to retrieve.

Request Lenses

getReportDefinition_reportId :: Lens' GetReportDefinition Text

ID of the report to retrieve.

Destructuring the Response

data GetReportDefinitionResponse

See: newGetReportDefinitionResponse smart constructor.

Constructors

GetReportDefinitionResponse'

Fields

httpStatus :: Int

The response's http status code.

reportId :: Text

ID of the report retrieved.

reportDescription :: Text

Description of the report.

reportFrequency :: ReportFrequency

Cadence used to generate the report.

format :: Format

Format of the generated report.

destinationS3Location :: S3Location

Amazon Simple Storage Service (Amazon S3) location where the report is uploaded.

createdAt :: POSIX

Timestamp (milliseconds) when this report definition was created.

lastUpdated :: POSIX

Timestamp (milliseconds) when this report definition was last updated.

newGetReportDefinitionResponse

Arguments

:: Int
$sel:httpStatus:GetReportDefinitionResponse'

-> Text
$sel:reportId:GetReportDefinition'

-> Text
$sel:reportDescription:GetReportDefinitionResponse'

-> ReportFrequency
$sel:reportFrequency:GetReportDefinitionResponse'

-> Format
$sel:format:GetReportDefinitionResponse'

-> S3Location
$sel:destinationS3Location:GetReportDefinitionResponse'

-> UTCTime
$sel:createdAt:GetReportDefinitionResponse'

-> UTCTime
$sel:lastUpdated:GetReportDefinitionResponse'

-> GetReportDefinitionResponse

Create a value of GetReportDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetReportDefinitionResponse', getReportDefinitionResponse_httpStatus - The response's http status code.

$sel:reportId:GetReportDefinition', getReportDefinitionResponse_reportId - ID of the report retrieved.

$sel:reportDescription:GetReportDefinitionResponse', getReportDefinitionResponse_reportDescription - Description of the report.

$sel:reportFrequency:GetReportDefinitionResponse', getReportDefinitionResponse_reportFrequency - Cadence used to generate the report.

$sel:format:GetReportDefinitionResponse', getReportDefinitionResponse_format - Format of the generated report.

$sel:destinationS3Location:GetReportDefinitionResponse', getReportDefinitionResponse_destinationS3Location - Amazon Simple Storage Service (Amazon S3) location where the report is uploaded.

$sel:createdAt:GetReportDefinitionResponse', getReportDefinitionResponse_createdAt - Timestamp (milliseconds) when this report definition was created.

$sel:lastUpdated:GetReportDefinitionResponse', getReportDefinitionResponse_lastUpdated - Timestamp (milliseconds) when this report definition was last updated.

Response Lenses

getReportDefinitionResponse_httpStatus :: Lens' GetReportDefinitionResponse Int

The response's http status code.

getReportDefinitionResponse_reportId :: Lens' GetReportDefinitionResponse Text

ID of the report retrieved.

getReportDefinitionResponse_reportDescription :: Lens' GetReportDefinitionResponse Text

Description of the report.

getReportDefinitionResponse_reportFrequency :: Lens' GetReportDefinitionResponse ReportFrequency

Cadence used to generate the report.

getReportDefinitionResponse_format :: Lens' GetReportDefinitionResponse Format

Format of the generated report.

getReportDefinitionResponse_destinationS3Location :: Lens' GetReportDefinitionResponse S3Location

Amazon Simple Storage Service (Amazon S3) location where the report is uploaded.

getReportDefinitionResponse_createdAt :: Lens' GetReportDefinitionResponse UTCTime

Timestamp (milliseconds) when this report definition was created.

getReportDefinitionResponse_lastUpdated :: Lens' GetReportDefinitionResponse UTCTime

Timestamp (milliseconds) when this report definition was last updated.