Network.AWS.ApplicationCostProfiler.ImportApplicationUsage

Description

Ingests application usage data from Amazon Simple Storage Service (Amazon S3).

The data must already exist in the S3 location. As part of the action, AWS Application Cost Profiler copies the object from your S3 bucket to an S3 bucket owned by Amazon for processing asynchronously.

Creating a Request

data ImportApplicationUsage

See: newImportApplicationUsage smart constructor.

Constructors

ImportApplicationUsage'

Fields

sourceS3Location :: SourceS3Location

Amazon S3 location to import application usage data from.

newImportApplicationUsage

Arguments

:: SourceS3Location
$sel:sourceS3Location:ImportApplicationUsage'

-> ImportApplicationUsage

Create a value of ImportApplicationUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceS3Location:ImportApplicationUsage', importApplicationUsage_sourceS3Location - Amazon S3 location to import application usage data from.

Request Lenses

importApplicationUsage_sourceS3Location :: Lens' ImportApplicationUsage SourceS3Location

Amazon S3 location to import application usage data from.

Destructuring the Response

data ImportApplicationUsageResponse

See: newImportApplicationUsageResponse smart constructor.

Constructors

ImportApplicationUsageResponse'

Fields

httpStatus :: Int

The response's http status code.

importId :: Text

ID of the import request.

newImportApplicationUsageResponse

Arguments

:: Int
$sel:httpStatus:ImportApplicationUsageResponse'

-> Text
$sel:importId:ImportApplicationUsageResponse'

-> ImportApplicationUsageResponse

Create a value of ImportApplicationUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ImportApplicationUsageResponse', importApplicationUsageResponse_httpStatus - The response's http status code.

$sel:importId:ImportApplicationUsageResponse', importApplicationUsageResponse_importId - ID of the import request.

Response Lenses

importApplicationUsageResponse_httpStatus :: Lens' ImportApplicationUsageResponse Int

The response's http status code.

importApplicationUsageResponse_importId :: Lens' ImportApplicationUsageResponse Text

ID of the import request.