Network.AWS.ApplicationInsights.DescribeComponent

Description

Describes a component and lists the resources that are grouped together in a component.

Creating a Request

data DescribeComponent

See: newDescribeComponent smart constructor.

Constructors

DescribeComponent'

Fields

resourceGroupName :: Text

The name of the resource group.

componentName :: Text

The name of the component.

newDescribeComponent

Arguments

:: Text
$sel:resourceGroupName:DescribeComponent'

-> Text
$sel:componentName:DescribeComponent'

-> DescribeComponent

Create a value of DescribeComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceGroupName:DescribeComponent', describeComponent_resourceGroupName - The name of the resource group.

$sel:componentName:DescribeComponent', describeComponent_componentName - The name of the component.

Request Lenses

describeComponent_resourceGroupName :: Lens' DescribeComponent Text

The name of the resource group.

describeComponent_componentName :: Lens' DescribeComponent Text

The name of the component.

Destructuring the Response

data DescribeComponentResponse

See: newDescribeComponentResponse smart constructor.

Constructors

DescribeComponentResponse'

Fields

applicationComponent :: Maybe ApplicationComponent

resourceList :: Maybe [Text]

The list of resource ARNs that belong to the component.

httpStatus :: Int

The response's http status code.

newDescribeComponentResponse

Arguments

:: Int
$sel:httpStatus:DescribeComponentResponse'

-> DescribeComponentResponse

Create a value of DescribeComponentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationComponent:DescribeComponentResponse', describeComponentResponse_applicationComponent - Undocumented member.

$sel:resourceList:DescribeComponentResponse', describeComponentResponse_resourceList - The list of resource ARNs that belong to the component.

$sel:httpStatus:DescribeComponentResponse', describeComponentResponse_httpStatus - The response's http status code.

Response Lenses

describeComponentResponse_applicationComponent :: Lens' DescribeComponentResponse (Maybe ApplicationComponent)

Undocumented member.

describeComponentResponse_resourceList :: Lens' DescribeComponentResponse (Maybe [Text])

The list of resource ARNs that belong to the component.

describeComponentResponse_httpStatus :: Lens' DescribeComponentResponse Int

The response's http status code.