Network.AWS.ApplicationInsights.DescribeObservation

Description

Describes an anomaly or error with the application.

Creating a Request

data DescribeObservation

See: newDescribeObservation smart constructor.

Constructors

DescribeObservation'

Fields

observationId :: Text

The ID of the observation.

newDescribeObservation

Arguments

:: Text
$sel:observationId:DescribeObservation'

-> DescribeObservation

Create a value of DescribeObservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:observationId:DescribeObservation', describeObservation_observationId - The ID of the observation.

Request Lenses

describeObservation_observationId :: Lens' DescribeObservation Text

The ID of the observation.

Destructuring the Response

data DescribeObservationResponse

See: newDescribeObservationResponse smart constructor.

Constructors

DescribeObservationResponse'

Fields

observation :: Maybe Observation

Information about the observation.

httpStatus :: Int

The response's http status code.

newDescribeObservationResponse

Arguments

:: Int
$sel:httpStatus:DescribeObservationResponse'

-> DescribeObservationResponse

Create a value of DescribeObservationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:observation:DescribeObservationResponse', describeObservationResponse_observation - Information about the observation.

$sel:httpStatus:DescribeObservationResponse', describeObservationResponse_httpStatus - The response's http status code.

Response Lenses

describeObservationResponse_observation :: Lens' DescribeObservationResponse (Maybe Observation)

Information about the observation.

describeObservationResponse_httpStatus :: Lens' DescribeObservationResponse Int

The response's http status code.