Network.AWS.ApplicationInsights.ListConfigurationHistory

Description

Lists the INFO, WARN, and ERROR events for periodic configuration updates performed by Application Insights. Examples of events represented are:

INFO: creating a new alarm or updating an alarm threshold.

WARN: alarm not created due to insufficient data points used to predict thresholds.

ERROR: alarm not created due to permission errors or exceeding quotas.

Creating a Request

data ListConfigurationHistory

See: newListConfigurationHistory smart constructor.

Constructors

ListConfigurationHistory'

Fields

resourceGroupName :: Maybe Text

Resource group to which the application belongs.

startTime :: Maybe POSIX

The start time of the event.

eventStatus :: Maybe ConfigurationEventStatus

The status of the configuration update event. Possible values include INFO, WARN, and ERROR.

nextToken :: Maybe Text

The NextToken value returned from a previous paginated ListConfigurationHistory request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

endTime :: Maybe POSIX

The end time of the event.

maxResults :: Maybe Natural

The maximum number of results returned by ListConfigurationHistory in paginated output. When this parameter is used, ListConfigurationHistory returns only MaxResults in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another ListConfigurationHistory request with the returned NextToken value. If this parameter is not used, then ListConfigurationHistory returns all results.

newListConfigurationHistory :: ListConfigurationHistory

Create a value of ListConfigurationHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceGroupName:ListConfigurationHistory', listConfigurationHistory_resourceGroupName - Resource group to which the application belongs.

$sel:startTime:ListConfigurationHistory', listConfigurationHistory_startTime - The start time of the event.

$sel:eventStatus:ListConfigurationHistory', listConfigurationHistory_eventStatus - The status of the configuration update event. Possible values include INFO, WARN, and ERROR.

$sel:nextToken:ListConfigurationHistory', listConfigurationHistory_nextToken - The NextToken value returned from a previous paginated ListConfigurationHistory request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

$sel:endTime:ListConfigurationHistory', listConfigurationHistory_endTime - The end time of the event.

$sel:maxResults:ListConfigurationHistory', listConfigurationHistory_maxResults - The maximum number of results returned by ListConfigurationHistory in paginated output. When this parameter is used, ListConfigurationHistory returns only MaxResults in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another ListConfigurationHistory request with the returned NextToken value. If this parameter is not used, then ListConfigurationHistory returns all results.

Request Lenses

listConfigurationHistory_resourceGroupName :: Lens' ListConfigurationHistory (Maybe Text)

Resource group to which the application belongs.

listConfigurationHistory_startTime :: Lens' ListConfigurationHistory (Maybe UTCTime)

The start time of the event.

listConfigurationHistory_eventStatus :: Lens' ListConfigurationHistory (Maybe ConfigurationEventStatus)

The status of the configuration update event. Possible values include INFO, WARN, and ERROR.

listConfigurationHistory_nextToken :: Lens' ListConfigurationHistory (Maybe Text)

The NextToken value returned from a previous paginated ListConfigurationHistory request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

listConfigurationHistory_endTime :: Lens' ListConfigurationHistory (Maybe UTCTime)

The end time of the event.

listConfigurationHistory_maxResults :: Lens' ListConfigurationHistory (Maybe Natural)

The maximum number of results returned by ListConfigurationHistory in paginated output. When this parameter is used, ListConfigurationHistory returns only MaxResults in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another ListConfigurationHistory request with the returned NextToken value. If this parameter is not used, then ListConfigurationHistory returns all results.

Destructuring the Response

data ListConfigurationHistoryResponse

See: newListConfigurationHistoryResponse smart constructor.

Constructors

ListConfigurationHistoryResponse'

Fields

nextToken :: Maybe Text

The NextToken value to include in a future ListConfigurationHistory request. When the results of a ListConfigurationHistory request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

eventList :: Maybe [ConfigurationEvent]

The list of configuration events and their corresponding details.

httpStatus :: Int

The response's http status code.

newListConfigurationHistoryResponse

Arguments

:: Int
$sel:httpStatus:ListConfigurationHistoryResponse'

-> ListConfigurationHistoryResponse

Create a value of ListConfigurationHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListConfigurationHistory', listConfigurationHistoryResponse_nextToken - The NextToken value to include in a future ListConfigurationHistory request. When the results of a ListConfigurationHistory request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:eventList:ListConfigurationHistoryResponse', listConfigurationHistoryResponse_eventList - The list of configuration events and their corresponding details.

$sel:httpStatus:ListConfigurationHistoryResponse', listConfigurationHistoryResponse_httpStatus - The response's http status code.

Response Lenses

listConfigurationHistoryResponse_nextToken :: Lens' ListConfigurationHistoryResponse (Maybe Text)

The NextToken value to include in a future ListConfigurationHistory request. When the results of a ListConfigurationHistory request exceed MaxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

listConfigurationHistoryResponse_eventList :: Lens' ListConfigurationHistoryResponse (Maybe [ConfigurationEvent])

The list of configuration events and their corresponding details.

listConfigurationHistoryResponse_httpStatus :: Lens' ListConfigurationHistoryResponse Int

The response's http status code.