Network.AWS.ApplicationInsights.ListLogPatternSets

Description

Lists the log pattern sets in the specific application.

Creating a Request

data ListLogPatternSets

See: newListLogPatternSets smart constructor.

Constructors

ListLogPatternSets'

Fields

nextToken :: Maybe Text

The token to request the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

resourceGroupName :: Text

The name of the resource group.

newListLogPatternSets

Arguments

:: Text
$sel:resourceGroupName:ListLogPatternSets'

-> ListLogPatternSets

Create a value of ListLogPatternSets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLogPatternSets', listLogPatternSets_nextToken - The token to request the next page of results.

$sel:maxResults:ListLogPatternSets', listLogPatternSets_maxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

$sel:resourceGroupName:ListLogPatternSets', listLogPatternSets_resourceGroupName - The name of the resource group.

Request Lenses

listLogPatternSets_nextToken :: Lens' ListLogPatternSets (Maybe Text)

The token to request the next page of results.

listLogPatternSets_maxResults :: Lens' ListLogPatternSets (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

listLogPatternSets_resourceGroupName :: Lens' ListLogPatternSets Text

The name of the resource group.

Destructuring the Response

data ListLogPatternSetsResponse

See: newListLogPatternSetsResponse smart constructor.

Constructors

ListLogPatternSetsResponse'

Fields

resourceGroupName :: Maybe Text

The name of the resource group.

nextToken :: Maybe Text

The token used to retrieve the next page of results. This value is null when there are no more results to return.

logPatternSets :: Maybe [Text]

The list of log pattern sets.

httpStatus :: Int

The response's http status code.

newListLogPatternSetsResponse

Arguments

:: Int
$sel:httpStatus:ListLogPatternSetsResponse'

-> ListLogPatternSetsResponse

Create a value of ListLogPatternSetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceGroupName:ListLogPatternSets', listLogPatternSetsResponse_resourceGroupName - The name of the resource group.

$sel:nextToken:ListLogPatternSets', listLogPatternSetsResponse_nextToken - The token used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:logPatternSets:ListLogPatternSetsResponse', listLogPatternSetsResponse_logPatternSets - The list of log pattern sets.

$sel:httpStatus:ListLogPatternSetsResponse', listLogPatternSetsResponse_httpStatus - The response's http status code.

Response Lenses

listLogPatternSetsResponse_resourceGroupName :: Lens' ListLogPatternSetsResponse (Maybe Text)

The name of the resource group.

listLogPatternSetsResponse_nextToken :: Lens' ListLogPatternSetsResponse (Maybe Text)

The token used to retrieve the next page of results. This value is null when there are no more results to return.

listLogPatternSetsResponse_logPatternSets :: Lens' ListLogPatternSetsResponse (Maybe [Text])

The list of log pattern sets.

listLogPatternSetsResponse_httpStatus :: Lens' ListLogPatternSetsResponse Int

The response's http status code.