Network.AWS.ApplicationInsights.Types

Description

Service Configuration

defaultService :: Service

API version 2018-11-25 of the Amazon CloudWatch Application Insights SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The parameter is not valid.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

User does not have permissions to perform this action.

_TagsAlreadyExistException :: AsError a => Getting (First ServiceError) a ServiceError

Tags are already registered for the specified application ARN.

_TooManyTagsException :: AsError a => Getting (First ServiceError) a ServiceError

The number of the provided tags is beyond the limit, or the number of total tags you are trying to attach to the specified resource exceeds the limit.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The server encountered an internal error and is unable to complete the request.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource does not exist in the customer account.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is not understood by the server.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is already created or in use.

CloudWatchEventSource

newtype CloudWatchEventSource

Constructors

CloudWatchEventSource'

Fields

fromCloudWatchEventSource :: Text

Bundled Patterns

pattern CloudWatchEventSource_CODE_DEPLOY :: CloudWatchEventSource

pattern CloudWatchEventSource_EC2 :: CloudWatchEventSource

pattern CloudWatchEventSource_HEALTH :: CloudWatchEventSource

pattern CloudWatchEventSource_RDS :: CloudWatchEventSource

ConfigurationEventResourceType

newtype ConfigurationEventResourceType

Constructors

ConfigurationEventResourceType'

Fields

fromConfigurationEventResourceType :: Text

Bundled Patterns

pattern ConfigurationEventResourceType_CLOUDFORMATION :: ConfigurationEventResourceType

pattern ConfigurationEventResourceType_CLOUDWATCH_ALARM :: ConfigurationEventResourceType

pattern ConfigurationEventResourceType_CLOUDWATCH_LOG :: ConfigurationEventResourceType

pattern ConfigurationEventResourceType_SSM_ASSOCIATION :: ConfigurationEventResourceType

ConfigurationEventStatus

newtype ConfigurationEventStatus

Constructors

ConfigurationEventStatus'

Fields

fromConfigurationEventStatus :: Text

Bundled Patterns

pattern ConfigurationEventStatus_ERROR :: ConfigurationEventStatus

pattern ConfigurationEventStatus_INFO :: ConfigurationEventStatus

pattern ConfigurationEventStatus_WARN :: ConfigurationEventStatus

FeedbackKey

newtype FeedbackKey

Constructors

FeedbackKey'

Fields

fromFeedbackKey :: Text

Bundled Patterns

pattern FeedbackKey_INSIGHTS_FEEDBACK :: FeedbackKey

FeedbackValue

newtype FeedbackValue

Constructors

FeedbackValue'

Fields

fromFeedbackValue :: Text

Bundled Patterns

pattern FeedbackValue_NOT_SPECIFIED :: FeedbackValue

pattern FeedbackValue_NOT_USEFUL :: FeedbackValue

pattern FeedbackValue_USEFUL :: FeedbackValue

LogFilter

newtype LogFilter

Constructors

LogFilter'

Fields

fromLogFilter :: Text

Bundled Patterns

pattern LogFilter_ERROR :: LogFilter

pattern LogFilter_INFO :: LogFilter

pattern LogFilter_WARN :: LogFilter

OsType

newtype OsType

Constructors

OsType'

Fields

fromOsType :: Text

Bundled Patterns

pattern OsType_LINUX :: OsType

pattern OsType_WINDOWS :: OsType

SeverityLevel

newtype SeverityLevel

Constructors

SeverityLevel'

Fields

fromSeverityLevel :: Text

Bundled Patterns

pattern SeverityLevel_High :: SeverityLevel

pattern SeverityLevel_Low :: SeverityLevel

pattern SeverityLevel_Medium :: SeverityLevel

Status

newtype Status

Constructors

Status'

Fields

fromStatus :: Text

Bundled Patterns

pattern Status_IGNORE :: Status

pattern Status_PENDING :: Status

pattern Status_RESOLVED :: Status

Tier

newtype Tier

Constructors

Tier'

Fields

fromTier :: Text

Bundled Patterns

pattern Tier_CUSTOM :: Tier

pattern Tier_DEFAULT :: Tier

pattern Tier_DOT_NET_CORE :: Tier

pattern Tier_DOT_NET_WEB :: Tier

pattern Tier_DOT_NET_WEB_TIER :: Tier

pattern Tier_DOT_NET_WORKER :: Tier

pattern Tier_JAVA_JMX :: Tier

pattern Tier_MYSQL :: Tier

pattern Tier_ORACLE :: Tier

pattern Tier_POSTGRESQL :: Tier

pattern Tier_SQL_SERVER :: Tier

pattern Tier_SQL_SERVER_ALWAYSON_AVAILABILITY_GROUP :: Tier

ApplicationComponent

data ApplicationComponent

Describes a standalone resource or similarly grouped resources that the application is made up of.

See: newApplicationComponent smart constructor.

Constructors

ApplicationComponent'

Fields

osType :: Maybe OsType

The operating system of the component.

resourceType :: Maybe Text

The resource type. Supported resource types include EC2 instances, Auto Scaling group, Classic ELB, Application ELB, and SQS Queue.

detectedWorkload :: Maybe (HashMap Tier (HashMap Text Text))

Workloads detected in the application component.

monitor :: Maybe Bool

Indicates whether the application component is monitored.

tier :: Maybe Tier

The stack tier of the application component.

componentName :: Maybe Text

The name of the component.

componentRemarks :: Maybe Text

If logging is supported for the resource type, indicates whether the component has configured logs to be monitored.

newApplicationComponent :: ApplicationComponent

Create a value of ApplicationComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:osType:ApplicationComponent', applicationComponent_osType - The operating system of the component.

$sel:resourceType:ApplicationComponent', applicationComponent_resourceType - The resource type. Supported resource types include EC2 instances, Auto Scaling group, Classic ELB, Application ELB, and SQS Queue.

$sel:detectedWorkload:ApplicationComponent', applicationComponent_detectedWorkload - Workloads detected in the application component.

$sel:monitor:ApplicationComponent', applicationComponent_monitor - Indicates whether the application component is monitored.

$sel:tier:ApplicationComponent', applicationComponent_tier - The stack tier of the application component.

$sel:componentName:ApplicationComponent', applicationComponent_componentName - The name of the component.

$sel:componentRemarks:ApplicationComponent', applicationComponent_componentRemarks - If logging is supported for the resource type, indicates whether the component has configured logs to be monitored.

applicationComponent_osType :: Lens' ApplicationComponent (Maybe OsType)

The operating system of the component.

applicationComponent_resourceType :: Lens' ApplicationComponent (Maybe Text)

The resource type. Supported resource types include EC2 instances, Auto Scaling group, Classic ELB, Application ELB, and SQS Queue.

applicationComponent_detectedWorkload :: Lens' ApplicationComponent (Maybe (HashMap Tier (HashMap Text Text)))

Workloads detected in the application component.

applicationComponent_monitor :: Lens' ApplicationComponent (Maybe Bool)

Indicates whether the application component is monitored.

applicationComponent_tier :: Lens' ApplicationComponent (Maybe Tier)

The stack tier of the application component.

applicationComponent_componentName :: Lens' ApplicationComponent (Maybe Text)

The name of the component.

applicationComponent_componentRemarks :: Lens' ApplicationComponent (Maybe Text)

If logging is supported for the resource type, indicates whether the component has configured logs to be monitored.

ApplicationInfo

data ApplicationInfo

Describes the status of the application.

See: newApplicationInfo smart constructor.

Constructors

ApplicationInfo'

Fields

resourceGroupName :: Maybe Text

The name of the resource group used for the application.

cWEMonitorEnabled :: Maybe Bool

Indicates whether Application Insights can listen to CloudWatch events for the application resources, such as instance terminated, failed deployment, and others.

opsItemSNSTopicArn :: Maybe Text

The SNS topic provided to Application Insights that is associated to the created opsItems to receive SNS notifications for opsItem updates.

lifeCycle :: Maybe Text

The lifecycle of the application.

opsCenterEnabled :: Maybe Bool

Indicates whether Application Insights will create opsItems for any problem detected by Application Insights for an application.

remarks :: Maybe Text

The issues on the user side that block Application Insights from successfully monitoring an application. Example remarks include:

“Configuring application, detected 1 Errors, 3 Warnings”

“Configuring application, detected 1 Unconfigured Components”

newApplicationInfo :: ApplicationInfo

Create a value of ApplicationInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceGroupName:ApplicationInfo', applicationInfo_resourceGroupName - The name of the resource group used for the application.

$sel:cWEMonitorEnabled:ApplicationInfo', applicationInfo_cWEMonitorEnabled - Indicates whether Application Insights can listen to CloudWatch events for the application resources, such as instance terminated, failed deployment, and others.

$sel:opsItemSNSTopicArn:ApplicationInfo', applicationInfo_opsItemSNSTopicArn - The SNS topic provided to Application Insights that is associated to the created opsItems to receive SNS notifications for opsItem updates.

$sel:lifeCycle:ApplicationInfo', applicationInfo_lifeCycle - The lifecycle of the application.

$sel:opsCenterEnabled:ApplicationInfo', applicationInfo_opsCenterEnabled - Indicates whether Application Insights will create opsItems for any problem detected by Application Insights for an application.

$sel:remarks:ApplicationInfo', applicationInfo_remarks - The issues on the user side that block Application Insights from successfully monitoring an application. Example remarks include:

“Configuring application, detected 1 Errors, 3 Warnings”

“Configuring application, detected 1 Unconfigured Components”

applicationInfo_resourceGroupName :: Lens' ApplicationInfo (Maybe Text)

The name of the resource group used for the application.

applicationInfo_cWEMonitorEnabled :: Lens' ApplicationInfo (Maybe Bool)

Indicates whether Application Insights can listen to CloudWatch events for the application resources, such as instance terminated, failed deployment, and others.

applicationInfo_opsItemSNSTopicArn :: Lens' ApplicationInfo (Maybe Text)

The SNS topic provided to Application Insights that is associated to the created opsItems to receive SNS notifications for opsItem updates.

applicationInfo_lifeCycle :: Lens' ApplicationInfo (Maybe Text)

The lifecycle of the application.

applicationInfo_opsCenterEnabled :: Lens' ApplicationInfo (Maybe Bool)

Indicates whether Application Insights will create opsItems for any problem detected by Application Insights for an application.

applicationInfo_remarks :: Lens' ApplicationInfo (Maybe Text)

The issues on the user side that block Application Insights from successfully monitoring an application. Example remarks include:

“Configuring application, detected 1 Errors, 3 Warnings”

“Configuring application, detected 1 Unconfigured Components”

ConfigurationEvent

data ConfigurationEvent

The event information.

See: newConfigurationEvent smart constructor.

Constructors

ConfigurationEvent'

Fields

monitoredResourceARN :: Maybe Text

The resource monitored by Application Insights.

eventStatus :: Maybe ConfigurationEventStatus

The status of the configuration update event. Possible values include INFO, WARN, and ERROR.

eventResourceName :: Maybe Text

The name of the resource Application Insights attempted to configure.

eventTime :: Maybe POSIX

The timestamp of the event.

eventDetail :: Maybe Text

The details of the event in plain text.

eventResourceType :: Maybe ConfigurationEventResourceType

The resource type that Application Insights attempted to configure, for example, CLOUDWATCH_ALARM.

newConfigurationEvent :: ConfigurationEvent

Create a value of ConfigurationEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monitoredResourceARN:ConfigurationEvent', configurationEvent_monitoredResourceARN - The resource monitored by Application Insights.

$sel:eventStatus:ConfigurationEvent', configurationEvent_eventStatus - The status of the configuration update event. Possible values include INFO, WARN, and ERROR.

$sel:eventResourceName:ConfigurationEvent', configurationEvent_eventResourceName - The name of the resource Application Insights attempted to configure.

$sel:eventTime:ConfigurationEvent', configurationEvent_eventTime - The timestamp of the event.

$sel:eventDetail:ConfigurationEvent', configurationEvent_eventDetail - The details of the event in plain text.

$sel:eventResourceType:ConfigurationEvent', configurationEvent_eventResourceType - The resource type that Application Insights attempted to configure, for example, CLOUDWATCH_ALARM.

configurationEvent_monitoredResourceARN :: Lens' ConfigurationEvent (Maybe Text)

The resource monitored by Application Insights.

configurationEvent_eventStatus :: Lens' ConfigurationEvent (Maybe ConfigurationEventStatus)

The status of the configuration update event. Possible values include INFO, WARN, and ERROR.

configurationEvent_eventResourceName :: Lens' ConfigurationEvent (Maybe Text)

The name of the resource Application Insights attempted to configure.

configurationEvent_eventTime :: Lens' ConfigurationEvent (Maybe UTCTime)

The timestamp of the event.

configurationEvent_eventDetail :: Lens' ConfigurationEvent (Maybe Text)

The details of the event in plain text.

configurationEvent_eventResourceType :: Lens' ConfigurationEvent (Maybe ConfigurationEventResourceType)

The resource type that Application Insights attempted to configure, for example, CLOUDWATCH_ALARM.

LogPattern

data LogPattern

An object that defines the log patterns that belongs to a LogPatternSet.

See: newLogPattern smart constructor.

Constructors

LogPattern'

Fields

pattern' :: Maybe Text

A regular expression that defines the log pattern. A log pattern can contain as many as 50 characters, and it cannot be empty. The pattern must be DFA compatible. Patterns that utilize forward lookahead or backreference constructions are not supported.

patternName :: Maybe Text

The name of the log pattern. A log pattern name can contain as many as 50 characters, and it cannot be empty. The characters can be Unicode letters, digits, or one of the following symbols: period, dash, underscore.

patternSetName :: Maybe Text

The name of the log pattern. A log pattern name can contain as many as 30 characters, and it cannot be empty. The characters can be Unicode letters, digits, or one of the following symbols: period, dash, underscore.

rank :: Maybe Int

Rank of the log pattern. Must be a value between 1 and 1,000,000. The patterns are sorted by rank, so we recommend that you set your highest priority patterns with the lowest rank. A pattern of rank 1 will be the first to get matched to a log line. A pattern of rank 1,000,000 will be last to get matched. When you configure custom log patterns from the console, a Low severity pattern translates to a 750,000 rank. A Medium severity pattern translates to a 500,000 rank. And a High severity pattern translates to a 250,000 rank. Rank values less than 1 or greater than 1,000,000 are reserved for AWS-provided patterns.

newLogPattern :: LogPattern

Create a value of LogPattern with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pattern':LogPattern', logPattern_pattern - A regular expression that defines the log pattern. A log pattern can contain as many as 50 characters, and it cannot be empty. The pattern must be DFA compatible. Patterns that utilize forward lookahead or backreference constructions are not supported.

$sel:patternName:LogPattern', logPattern_patternName - The name of the log pattern. A log pattern name can contain as many as 50 characters, and it cannot be empty. The characters can be Unicode letters, digits, or one of the following symbols: period, dash, underscore.

$sel:patternSetName:LogPattern', logPattern_patternSetName - The name of the log pattern. A log pattern name can contain as many as 30 characters, and it cannot be empty. The characters can be Unicode letters, digits, or one of the following symbols: period, dash, underscore.

$sel:rank:LogPattern', logPattern_rank - Rank of the log pattern. Must be a value between 1 and 1,000,000. The patterns are sorted by rank, so we recommend that you set your highest priority patterns with the lowest rank. A pattern of rank 1 will be the first to get matched to a log line. A pattern of rank 1,000,000 will be last to get matched. When you configure custom log patterns from the console, a Low severity pattern translates to a 750,000 rank. A Medium severity pattern translates to a 500,000 rank. And a High severity pattern translates to a 250,000 rank. Rank values less than 1 or greater than 1,000,000 are reserved for AWS-provided patterns.

logPattern_pattern :: Lens' LogPattern (Maybe Text)

A regular expression that defines the log pattern. A log pattern can contain as many as 50 characters, and it cannot be empty. The pattern must be DFA compatible. Patterns that utilize forward lookahead or backreference constructions are not supported.

logPattern_patternName :: Lens' LogPattern (Maybe Text)

The name of the log pattern. A log pattern name can contain as many as 50 characters, and it cannot be empty. The characters can be Unicode letters, digits, or one of the following symbols: period, dash, underscore.

logPattern_patternSetName :: Lens' LogPattern (Maybe Text)

The name of the log pattern. A log pattern name can contain as many as 30 characters, and it cannot be empty. The characters can be Unicode letters, digits, or one of the following symbols: period, dash, underscore.

logPattern_rank :: Lens' LogPattern (Maybe Int)

Rank of the log pattern. Must be a value between 1 and 1,000,000. The patterns are sorted by rank, so we recommend that you set your highest priority patterns with the lowest rank. A pattern of rank 1 will be the first to get matched to a log line. A pattern of rank 1,000,000 will be last to get matched. When you configure custom log patterns from the console, a Low severity pattern translates to a 750,000 rank. A Medium severity pattern translates to a 500,000 rank. And a High severity pattern translates to a 250,000 rank. Rank values less than 1 or greater than 1,000,000 are reserved for AWS-provided patterns.

Observation

data Observation

Describes an anomaly or error with the application.

See: newObservation smart constructor.

Constructors

Observation'

Fields

codeDeployApplication :: Maybe Text

The CodeDeploy application to which the deployment belongs.

rdsEventMessage :: Maybe Text

The message of an RDS event.

codeDeployDeploymentId :: Maybe Text

The deployment ID of the CodeDeploy-based observation related to the detected problem.

startTime :: Maybe POSIX

The time when the observation was first detected, in epoch seconds.

sourceType :: Maybe Text

The source type of the observation.

sourceARN :: Maybe Text

The source resource ARN of the observation.

xRayRequestAverageLatency :: Maybe Integer

The X-Ray node request average latency for this node.

statesStatus :: Maybe Text

The status of the step function-related observation.

codeDeployDeploymentGroup :: Maybe Text

The deployment group to which the CodeDeploy deployment belongs.

healthEventTypeCategory :: Maybe Text

The category of the AWS Health event, such as issue.

xRayRequestCount :: Maybe Int

The X-Ray request count for this node.

s3EventName :: Maybe Text

The name of the S3 CloudWatch Event-based observation.

metricName :: Maybe Text

The name of the observation metric.

ec2State :: Maybe Text

The state of the instance, such as STOPPING or TERMINATING.

logGroup :: Maybe Text

The log group name.

value :: Maybe Double

The value of the source observation metric.

healthEventDescription :: Maybe Text

The description of the AWS Health event provided by the service, such as Amazon EC2.

cloudWatchEventSource :: Maybe CloudWatchEventSource

The source of the CloudWatch Event.

codeDeployState :: Maybe Text

The status of the CodeDeploy deployment, for example SUCCESS or FAILURE.

xRayErrorPercent :: Maybe Int

The X-Ray request error percentage for this node.

statesArn :: Maybe Text

The Amazon Resource Name (ARN) of the step function-based observation.

cloudWatchEventId :: Maybe Text

The ID of the CloudWatch Event-based observation related to the detected problem.

logText :: Maybe Text

The log text of the observation.

logFilter :: Maybe LogFilter

The log filter of the observation.

metricNamespace :: Maybe Text

The namespace of the observation metric.

rdsEventCategories :: Maybe Text

The category of an RDS event.

xRayNodeType :: Maybe Text

The type of the X-Ray node.

endTime :: Maybe POSIX

The time when the observation ended, in epoch seconds.

statesInput :: Maybe Text

The input to the step function-based observation.

xRayNodeName :: Maybe Text

The name of the X-Ray node.

id :: Maybe Text

The ID of the observation type.

healthEventArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS Health Event-based observation.

healthEventTypeCode :: Maybe Text

The type of the AWS Health event, for example, AWS_EC2_POWER_CONNECTIVITY_ISSUE.

ebsResult :: Maybe Text

The result of an EBS CloudWatch event, such as failed or succeeded.

cloudWatchEventDetailType :: Maybe Text

The detail type of the CloudWatch Event-based observation, for example, EC2 Instance State-change Notification.

codeDeployInstanceGroupId :: Maybe Text

The instance group to which the CodeDeploy instance belongs.

ebsCause :: Maybe Text

The cause of an EBS CloudWatch event.

ebsEvent :: Maybe Text

The type of EBS CloudWatch event, such as createVolume, deleteVolume or attachVolume.

ebsRequestId :: Maybe Text

The request ID of an EBS CloudWatch event.

xRayFaultPercent :: Maybe Int

The X-Ray request fault percentage for this node.

statesExecutionArn :: Maybe Text

The Amazon Resource Name (ARN) of the step function execution-based observation.

lineTime :: Maybe POSIX

The timestamp in the CloudWatch Logs that specifies when the matched line occurred.

unit :: Maybe Text

The unit of the source observation metric.

xRayThrottlePercent :: Maybe Int

The X-Ray request throttle percentage for this node.

healthService :: Maybe Text

The service to which the AWS Health Event belongs, such as EC2.

newObservation :: Observation

Create a value of Observation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeDeployApplication:Observation', observation_codeDeployApplication - The CodeDeploy application to which the deployment belongs.

$sel:rdsEventMessage:Observation', observation_rdsEventMessage - The message of an RDS event.

$sel:codeDeployDeploymentId:Observation', observation_codeDeployDeploymentId - The deployment ID of the CodeDeploy-based observation related to the detected problem.

$sel:startTime:Observation', observation_startTime - The time when the observation was first detected, in epoch seconds.

$sel:sourceType:Observation', observation_sourceType - The source type of the observation.

$sel:sourceARN:Observation', observation_sourceARN - The source resource ARN of the observation.

$sel:xRayRequestAverageLatency:Observation', observation_xRayRequestAverageLatency - The X-Ray node request average latency for this node.

$sel:statesStatus:Observation', observation_statesStatus - The status of the step function-related observation.

$sel:codeDeployDeploymentGroup:Observation', observation_codeDeployDeploymentGroup - The deployment group to which the CodeDeploy deployment belongs.

$sel:healthEventTypeCategory:Observation', observation_healthEventTypeCategory - The category of the AWS Health event, such as issue.

$sel:xRayRequestCount:Observation', observation_xRayRequestCount - The X-Ray request count for this node.

$sel:s3EventName:Observation', observation_s3EventName - The name of the S3 CloudWatch Event-based observation.

$sel:metricName:Observation', observation_metricName - The name of the observation metric.

$sel:ec2State:Observation', observation_ec2State - The state of the instance, such as STOPPING or TERMINATING.

$sel:logGroup:Observation', observation_logGroup - The log group name.

$sel:value:Observation', observation_value - The value of the source observation metric.

$sel:healthEventDescription:Observation', observation_healthEventDescription - The description of the AWS Health event provided by the service, such as Amazon EC2.

$sel:cloudWatchEventSource:Observation', observation_cloudWatchEventSource - The source of the CloudWatch Event.

$sel:codeDeployState:Observation', observation_codeDeployState - The status of the CodeDeploy deployment, for example SUCCESS or FAILURE.

$sel:xRayErrorPercent:Observation', observation_xRayErrorPercent - The X-Ray request error percentage for this node.

$sel:statesArn:Observation', observation_statesArn - The Amazon Resource Name (ARN) of the step function-based observation.

$sel:cloudWatchEventId:Observation', observation_cloudWatchEventId - The ID of the CloudWatch Event-based observation related to the detected problem.

$sel:logText:Observation', observation_logText - The log text of the observation.

$sel:logFilter:Observation', observation_logFilter - The log filter of the observation.

$sel:metricNamespace:Observation', observation_metricNamespace - The namespace of the observation metric.

$sel:rdsEventCategories:Observation', observation_rdsEventCategories - The category of an RDS event.

$sel:xRayNodeType:Observation', observation_xRayNodeType - The type of the X-Ray node.

$sel:endTime:Observation', observation_endTime - The time when the observation ended, in epoch seconds.

$sel:statesInput:Observation', observation_statesInput - The input to the step function-based observation.

$sel:xRayNodeName:Observation', observation_xRayNodeName - The name of the X-Ray node.

$sel:id:Observation', observation_id - The ID of the observation type.

$sel:healthEventArn:Observation', observation_healthEventArn - The Amazon Resource Name (ARN) of the AWS Health Event-based observation.

$sel:healthEventTypeCode:Observation', observation_healthEventTypeCode - The type of the AWS Health event, for example, AWS_EC2_POWER_CONNECTIVITY_ISSUE.

$sel:ebsResult:Observation', observation_ebsResult - The result of an EBS CloudWatch event, such as failed or succeeded.

$sel:cloudWatchEventDetailType:Observation', observation_cloudWatchEventDetailType - The detail type of the CloudWatch Event-based observation, for example, EC2 Instance State-change Notification.

$sel:codeDeployInstanceGroupId:Observation', observation_codeDeployInstanceGroupId - The instance group to which the CodeDeploy instance belongs.

$sel:ebsCause:Observation', observation_ebsCause - The cause of an EBS CloudWatch event.

$sel:ebsEvent:Observation', observation_ebsEvent - The type of EBS CloudWatch event, such as createVolume, deleteVolume or attachVolume.

$sel:ebsRequestId:Observation', observation_ebsRequestId - The request ID of an EBS CloudWatch event.

$sel:xRayFaultPercent:Observation', observation_xRayFaultPercent - The X-Ray request fault percentage for this node.

$sel:statesExecutionArn:Observation', observation_statesExecutionArn - The Amazon Resource Name (ARN) of the step function execution-based observation.

$sel:lineTime:Observation', observation_lineTime - The timestamp in the CloudWatch Logs that specifies when the matched line occurred.

$sel:unit:Observation', observation_unit - The unit of the source observation metric.

$sel:xRayThrottlePercent:Observation', observation_xRayThrottlePercent - The X-Ray request throttle percentage for this node.

$sel:healthService:Observation', observation_healthService - The service to which the AWS Health Event belongs, such as EC2.

observation_codeDeployApplication :: Lens' Observation (Maybe Text)

The CodeDeploy application to which the deployment belongs.

observation_rdsEventMessage :: Lens' Observation (Maybe Text)

The message of an RDS event.

observation_codeDeployDeploymentId :: Lens' Observation (Maybe Text)

The deployment ID of the CodeDeploy-based observation related to the detected problem.

observation_startTime :: Lens' Observation (Maybe UTCTime)

The time when the observation was first detected, in epoch seconds.

observation_sourceType :: Lens' Observation (Maybe Text)

The source type of the observation.

observation_sourceARN :: Lens' Observation (Maybe Text)

The source resource ARN of the observation.

observation_xRayRequestAverageLatency :: Lens' Observation (Maybe Integer)

The X-Ray node request average latency for this node.

observation_statesStatus :: Lens' Observation (Maybe Text)

The status of the step function-related observation.

observation_codeDeployDeploymentGroup :: Lens' Observation (Maybe Text)

The deployment group to which the CodeDeploy deployment belongs.

observation_healthEventTypeCategory :: Lens' Observation (Maybe Text)

The category of the AWS Health event, such as issue.

observation_xRayRequestCount :: Lens' Observation (Maybe Int)

The X-Ray request count for this node.

observation_s3EventName :: Lens' Observation (Maybe Text)

The name of the S3 CloudWatch Event-based observation.

observation_metricName :: Lens' Observation (Maybe Text)

The name of the observation metric.

observation_ec2State :: Lens' Observation (Maybe Text)

The state of the instance, such as STOPPING or TERMINATING.

observation_logGroup :: Lens' Observation (Maybe Text)

The log group name.

observation_value :: Lens' Observation (Maybe Double)

The value of the source observation metric.

observation_healthEventDescription :: Lens' Observation (Maybe Text)

The description of the AWS Health event provided by the service, such as Amazon EC2.

observation_cloudWatchEventSource :: Lens' Observation (Maybe CloudWatchEventSource)

The source of the CloudWatch Event.

observation_codeDeployState :: Lens' Observation (Maybe Text)

The status of the CodeDeploy deployment, for example SUCCESS or FAILURE.

observation_xRayErrorPercent :: Lens' Observation (Maybe Int)

The X-Ray request error percentage for this node.

observation_statesArn :: Lens' Observation (Maybe Text)

The Amazon Resource Name (ARN) of the step function-based observation.

observation_cloudWatchEventId :: Lens' Observation (Maybe Text)

The ID of the CloudWatch Event-based observation related to the detected problem.

observation_logText :: Lens' Observation (Maybe Text)

The log text of the observation.

observation_logFilter :: Lens' Observation (Maybe LogFilter)

The log filter of the observation.

observation_metricNamespace :: Lens' Observation (Maybe Text)

The namespace of the observation metric.

observation_rdsEventCategories :: Lens' Observation (Maybe Text)

The category of an RDS event.

observation_xRayNodeType :: Lens' Observation (Maybe Text)

The type of the X-Ray node.

observation_endTime :: Lens' Observation (Maybe UTCTime)

The time when the observation ended, in epoch seconds.

observation_statesInput :: Lens' Observation (Maybe Text)

The input to the step function-based observation.

observation_xRayNodeName :: Lens' Observation (Maybe Text)

The name of the X-Ray node.

observation_id :: Lens' Observation (Maybe Text)

The ID of the observation type.

observation_healthEventArn :: Lens' Observation (Maybe Text)

The Amazon Resource Name (ARN) of the AWS Health Event-based observation.

observation_healthEventTypeCode :: Lens' Observation (Maybe Text)

The type of the AWS Health event, for example, AWS_EC2_POWER_CONNECTIVITY_ISSUE.

observation_ebsResult :: Lens' Observation (Maybe Text)

The result of an EBS CloudWatch event, such as failed or succeeded.

observation_cloudWatchEventDetailType :: Lens' Observation (Maybe Text)

The detail type of the CloudWatch Event-based observation, for example, EC2 Instance State-change Notification.

observation_codeDeployInstanceGroupId :: Lens' Observation (Maybe Text)

The instance group to which the CodeDeploy instance belongs.

observation_ebsCause :: Lens' Observation (Maybe Text)

The cause of an EBS CloudWatch event.

observation_ebsEvent :: Lens' Observation (Maybe Text)

The type of EBS CloudWatch event, such as createVolume, deleteVolume or attachVolume.

observation_ebsRequestId :: Lens' Observation (Maybe Text)

The request ID of an EBS CloudWatch event.

observation_xRayFaultPercent :: Lens' Observation (Maybe Int)

The X-Ray request fault percentage for this node.

observation_statesExecutionArn :: Lens' Observation (Maybe Text)

The Amazon Resource Name (ARN) of the step function execution-based observation.

observation_lineTime :: Lens' Observation (Maybe UTCTime)

The timestamp in the CloudWatch Logs that specifies when the matched line occurred.

observation_unit :: Lens' Observation (Maybe Text)

The unit of the source observation metric.

observation_xRayThrottlePercent :: Lens' Observation (Maybe Int)

The X-Ray request throttle percentage for this node.

observation_healthService :: Lens' Observation (Maybe Text)

The service to which the AWS Health Event belongs, such as EC2.

Problem

data Problem

Describes a problem that is detected by correlating observations.

See: newProblem smart constructor.

Constructors

Problem'

Fields

status :: Maybe Status

The status of the problem.

resourceGroupName :: Maybe Text

The name of the resource group affected by the problem.

startTime :: Maybe POSIX

The time when the problem started, in epoch seconds.

insights :: Maybe Text

A detailed analysis of the problem using machine learning.

endTime :: Maybe POSIX

The time when the problem ended, in epoch seconds.

id :: Maybe Text

The ID of the problem.

severityLevel :: Maybe SeverityLevel

A measure of the level of impact of the problem.

title :: Maybe Text

The name of the problem.

affectedResource :: Maybe Text

The resource affected by the problem.

feedback :: Maybe (HashMap FeedbackKey FeedbackValue)

Feedback provided by the user about the problem.

newProblem :: Problem

Create a value of Problem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Problem', problem_status - The status of the problem.

$sel:resourceGroupName:Problem', problem_resourceGroupName - The name of the resource group affected by the problem.

$sel:startTime:Problem', problem_startTime - The time when the problem started, in epoch seconds.

$sel:insights:Problem', problem_insights - A detailed analysis of the problem using machine learning.

$sel:endTime:Problem', problem_endTime - The time when the problem ended, in epoch seconds.

$sel:id:Problem', problem_id - The ID of the problem.

$sel:severityLevel:Problem', problem_severityLevel - A measure of the level of impact of the problem.

$sel:title:Problem', problem_title - The name of the problem.

$sel:affectedResource:Problem', problem_affectedResource - The resource affected by the problem.

$sel:feedback:Problem', problem_feedback - Feedback provided by the user about the problem.

problem_status :: Lens' Problem (Maybe Status)

The status of the problem.

problem_resourceGroupName :: Lens' Problem (Maybe Text)

The name of the resource group affected by the problem.

problem_startTime :: Lens' Problem (Maybe UTCTime)

The time when the problem started, in epoch seconds.

problem_insights :: Lens' Problem (Maybe Text)

A detailed analysis of the problem using machine learning.

problem_endTime :: Lens' Problem (Maybe UTCTime)

The time when the problem ended, in epoch seconds.

problem_id :: Lens' Problem (Maybe Text)

The ID of the problem.

problem_severityLevel :: Lens' Problem (Maybe SeverityLevel)

A measure of the level of impact of the problem.

problem_title :: Lens' Problem (Maybe Text)

The name of the problem.

problem_affectedResource :: Lens' Problem (Maybe Text)

The resource affected by the problem.

problem_feedback :: Lens' Problem (Maybe (HashMap FeedbackKey FeedbackValue))

Feedback provided by the user about the problem.

RelatedObservations

data RelatedObservations

Describes observations related to the problem.

See: newRelatedObservations smart constructor.

Constructors

RelatedObservations'

Fields

observationList :: Maybe [Observation]

The list of observations related to the problem.

newRelatedObservations :: RelatedObservations

Create a value of RelatedObservations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:observationList:RelatedObservations', relatedObservations_observationList - The list of observations related to the problem.

relatedObservations_observationList :: Lens' RelatedObservations (Maybe [Observation])

The list of observations related to the problem.

Tag

data Tag

An object that defines the tags associated with an application. A tag is a label that you optionally define and associate with an application. Tags can help you categorize and manage resources in different ways, such as by purpose, owner, environment, or other criteria.

Each tag consists of a required tag key and an associated tag value, both of which you define. A tag key is a general label that acts as a category for a more specific tag value. A tag value acts as a descriptor within a tag key. A tag key can contain as many as 128 characters. A tag value can contain as many as 256 characters. The characters can be Unicode letters, digits, white space, or one of the following symbols: _ . : / = + -. The following additional restrictions apply to tags:

Tag keys and values are case sensitive.

For each associated resource, each tag key must be unique and it can have only one value.

The aws: prefix is reserved for use by AWS; you can’t use it in any tag keys or values that you define. In addition, you can't edit or remove tag keys or values that use this prefix.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

One part of a key-value pair that defines a tag. The maximum length of a tag key is 128 characters. The minimum length is 1 character.

value :: Text

The optional part of a key-value pair that defines a tag. The maximum length of a tag value is 256 characters. The minimum length is 0 characters. If you don't want an application to have a specific tag value, don't specify a value for this parameter.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - One part of a key-value pair that defines a tag. The maximum length of a tag key is 128 characters. The minimum length is 1 character.

$sel:value:Tag', tag_value - The optional part of a key-value pair that defines a tag. The maximum length of a tag value is 256 characters. The minimum length is 0 characters. If you don't want an application to have a specific tag value, don't specify a value for this parameter.

tag_key :: Lens' Tag Text

One part of a key-value pair that defines a tag. The maximum length of a tag key is 128 characters. The minimum length is 1 character.

tag_value :: Lens' Tag Text

The optional part of a key-value pair that defines a tag. The maximum length of a tag value is 256 characters. The minimum length is 0 characters. If you don't want an application to have a specific tag value, don't specify a value for this parameter.