Network.AWS.ApplicationInsights.Types.Observation

Description

Documentation

data Observation

Describes an anomaly or error with the application.

See: newObservation smart constructor.

Constructors

Observation'

Fields

codeDeployApplication :: Maybe Text

The CodeDeploy application to which the deployment belongs.

rdsEventMessage :: Maybe Text

The message of an RDS event.

codeDeployDeploymentId :: Maybe Text

The deployment ID of the CodeDeploy-based observation related to the detected problem.

startTime :: Maybe POSIX

The time when the observation was first detected, in epoch seconds.

sourceType :: Maybe Text

The source type of the observation.

sourceARN :: Maybe Text

The source resource ARN of the observation.

xRayRequestAverageLatency :: Maybe Integer

The X-Ray node request average latency for this node.

statesStatus :: Maybe Text

The status of the step function-related observation.

codeDeployDeploymentGroup :: Maybe Text

The deployment group to which the CodeDeploy deployment belongs.

healthEventTypeCategory :: Maybe Text

The category of the AWS Health event, such as issue.

xRayRequestCount :: Maybe Int

The X-Ray request count for this node.

s3EventName :: Maybe Text

The name of the S3 CloudWatch Event-based observation.

metricName :: Maybe Text

The name of the observation metric.

ec2State :: Maybe Text

The state of the instance, such as STOPPING or TERMINATING.

logGroup :: Maybe Text

The log group name.

value :: Maybe Double

The value of the source observation metric.

healthEventDescription :: Maybe Text

The description of the AWS Health event provided by the service, such as Amazon EC2.

cloudWatchEventSource :: Maybe CloudWatchEventSource

The source of the CloudWatch Event.

codeDeployState :: Maybe Text

The status of the CodeDeploy deployment, for example SUCCESS or FAILURE.

xRayErrorPercent :: Maybe Int

The X-Ray request error percentage for this node.

statesArn :: Maybe Text

The Amazon Resource Name (ARN) of the step function-based observation.

cloudWatchEventId :: Maybe Text

The ID of the CloudWatch Event-based observation related to the detected problem.

logText :: Maybe Text

The log text of the observation.

logFilter :: Maybe LogFilter

The log filter of the observation.

metricNamespace :: Maybe Text

The namespace of the observation metric.

rdsEventCategories :: Maybe Text

The category of an RDS event.

xRayNodeType :: Maybe Text

The type of the X-Ray node.

endTime :: Maybe POSIX

The time when the observation ended, in epoch seconds.

statesInput :: Maybe Text

The input to the step function-based observation.

xRayNodeName :: Maybe Text

The name of the X-Ray node.

id :: Maybe Text

The ID of the observation type.

healthEventArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS Health Event-based observation.

healthEventTypeCode :: Maybe Text

The type of the AWS Health event, for example, AWS_EC2_POWER_CONNECTIVITY_ISSUE.

ebsResult :: Maybe Text

The result of an EBS CloudWatch event, such as failed or succeeded.

cloudWatchEventDetailType :: Maybe Text

The detail type of the CloudWatch Event-based observation, for example, EC2 Instance State-change Notification.

codeDeployInstanceGroupId :: Maybe Text

The instance group to which the CodeDeploy instance belongs.

ebsCause :: Maybe Text

The cause of an EBS CloudWatch event.

ebsEvent :: Maybe Text

The type of EBS CloudWatch event, such as createVolume, deleteVolume or attachVolume.

ebsRequestId :: Maybe Text

The request ID of an EBS CloudWatch event.

xRayFaultPercent :: Maybe Int

The X-Ray request fault percentage for this node.

statesExecutionArn :: Maybe Text

The Amazon Resource Name (ARN) of the step function execution-based observation.

lineTime :: Maybe POSIX

The timestamp in the CloudWatch Logs that specifies when the matched line occurred.

unit :: Maybe Text

The unit of the source observation metric.

xRayThrottlePercent :: Maybe Int

The X-Ray request throttle percentage for this node.

healthService :: Maybe Text

The service to which the AWS Health Event belongs, such as EC2.

newObservation :: Observation

Create a value of Observation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeDeployApplication:Observation', observation_codeDeployApplication - The CodeDeploy application to which the deployment belongs.

$sel:rdsEventMessage:Observation', observation_rdsEventMessage - The message of an RDS event.

$sel:codeDeployDeploymentId:Observation', observation_codeDeployDeploymentId - The deployment ID of the CodeDeploy-based observation related to the detected problem.

$sel:startTime:Observation', observation_startTime - The time when the observation was first detected, in epoch seconds.

$sel:sourceType:Observation', observation_sourceType - The source type of the observation.

$sel:sourceARN:Observation', observation_sourceARN - The source resource ARN of the observation.

$sel:xRayRequestAverageLatency:Observation', observation_xRayRequestAverageLatency - The X-Ray node request average latency for this node.

$sel:statesStatus:Observation', observation_statesStatus - The status of the step function-related observation.

$sel:codeDeployDeploymentGroup:Observation', observation_codeDeployDeploymentGroup - The deployment group to which the CodeDeploy deployment belongs.

$sel:healthEventTypeCategory:Observation', observation_healthEventTypeCategory - The category of the AWS Health event, such as issue.

$sel:xRayRequestCount:Observation', observation_xRayRequestCount - The X-Ray request count for this node.

$sel:s3EventName:Observation', observation_s3EventName - The name of the S3 CloudWatch Event-based observation.

$sel:metricName:Observation', observation_metricName - The name of the observation metric.

$sel:ec2State:Observation', observation_ec2State - The state of the instance, such as STOPPING or TERMINATING.

$sel:logGroup:Observation', observation_logGroup - The log group name.

$sel:value:Observation', observation_value - The value of the source observation metric.

$sel:healthEventDescription:Observation', observation_healthEventDescription - The description of the AWS Health event provided by the service, such as Amazon EC2.

$sel:cloudWatchEventSource:Observation', observation_cloudWatchEventSource - The source of the CloudWatch Event.

$sel:codeDeployState:Observation', observation_codeDeployState - The status of the CodeDeploy deployment, for example SUCCESS or FAILURE.

$sel:xRayErrorPercent:Observation', observation_xRayErrorPercent - The X-Ray request error percentage for this node.

$sel:statesArn:Observation', observation_statesArn - The Amazon Resource Name (ARN) of the step function-based observation.

$sel:cloudWatchEventId:Observation', observation_cloudWatchEventId - The ID of the CloudWatch Event-based observation related to the detected problem.

$sel:logText:Observation', observation_logText - The log text of the observation.

$sel:logFilter:Observation', observation_logFilter - The log filter of the observation.

$sel:metricNamespace:Observation', observation_metricNamespace - The namespace of the observation metric.

$sel:rdsEventCategories:Observation', observation_rdsEventCategories - The category of an RDS event.

$sel:xRayNodeType:Observation', observation_xRayNodeType - The type of the X-Ray node.

$sel:endTime:Observation', observation_endTime - The time when the observation ended, in epoch seconds.

$sel:statesInput:Observation', observation_statesInput - The input to the step function-based observation.

$sel:xRayNodeName:Observation', observation_xRayNodeName - The name of the X-Ray node.

$sel:id:Observation', observation_id - The ID of the observation type.

$sel:healthEventArn:Observation', observation_healthEventArn - The Amazon Resource Name (ARN) of the AWS Health Event-based observation.

$sel:healthEventTypeCode:Observation', observation_healthEventTypeCode - The type of the AWS Health event, for example, AWS_EC2_POWER_CONNECTIVITY_ISSUE.

$sel:ebsResult:Observation', observation_ebsResult - The result of an EBS CloudWatch event, such as failed or succeeded.

$sel:cloudWatchEventDetailType:Observation', observation_cloudWatchEventDetailType - The detail type of the CloudWatch Event-based observation, for example, EC2 Instance State-change Notification.

$sel:codeDeployInstanceGroupId:Observation', observation_codeDeployInstanceGroupId - The instance group to which the CodeDeploy instance belongs.

$sel:ebsCause:Observation', observation_ebsCause - The cause of an EBS CloudWatch event.

$sel:ebsEvent:Observation', observation_ebsEvent - The type of EBS CloudWatch event, such as createVolume, deleteVolume or attachVolume.

$sel:ebsRequestId:Observation', observation_ebsRequestId - The request ID of an EBS CloudWatch event.

$sel:xRayFaultPercent:Observation', observation_xRayFaultPercent - The X-Ray request fault percentage for this node.

$sel:statesExecutionArn:Observation', observation_statesExecutionArn - The Amazon Resource Name (ARN) of the step function execution-based observation.

$sel:lineTime:Observation', observation_lineTime - The timestamp in the CloudWatch Logs that specifies when the matched line occurred.

$sel:unit:Observation', observation_unit - The unit of the source observation metric.

$sel:xRayThrottlePercent:Observation', observation_xRayThrottlePercent - The X-Ray request throttle percentage for this node.

$sel:healthService:Observation', observation_healthService - The service to which the AWS Health Event belongs, such as EC2.

observation_codeDeployApplication :: Lens' Observation (Maybe Text)

The CodeDeploy application to which the deployment belongs.

observation_rdsEventMessage :: Lens' Observation (Maybe Text)

The message of an RDS event.

observation_codeDeployDeploymentId :: Lens' Observation (Maybe Text)

The deployment ID of the CodeDeploy-based observation related to the detected problem.

observation_startTime :: Lens' Observation (Maybe UTCTime)

The time when the observation was first detected, in epoch seconds.

observation_sourceType :: Lens' Observation (Maybe Text)

The source type of the observation.

observation_sourceARN :: Lens' Observation (Maybe Text)

The source resource ARN of the observation.

observation_xRayRequestAverageLatency :: Lens' Observation (Maybe Integer)

The X-Ray node request average latency for this node.

observation_statesStatus :: Lens' Observation (Maybe Text)

The status of the step function-related observation.

observation_codeDeployDeploymentGroup :: Lens' Observation (Maybe Text)

The deployment group to which the CodeDeploy deployment belongs.

observation_healthEventTypeCategory :: Lens' Observation (Maybe Text)

The category of the AWS Health event, such as issue.

observation_xRayRequestCount :: Lens' Observation (Maybe Int)

The X-Ray request count for this node.

observation_s3EventName :: Lens' Observation (Maybe Text)

The name of the S3 CloudWatch Event-based observation.

observation_metricName :: Lens' Observation (Maybe Text)

The name of the observation metric.

observation_ec2State :: Lens' Observation (Maybe Text)

The state of the instance, such as STOPPING or TERMINATING.

observation_logGroup :: Lens' Observation (Maybe Text)

The log group name.

observation_value :: Lens' Observation (Maybe Double)

The value of the source observation metric.

observation_healthEventDescription :: Lens' Observation (Maybe Text)

The description of the AWS Health event provided by the service, such as Amazon EC2.

observation_cloudWatchEventSource :: Lens' Observation (Maybe CloudWatchEventSource)

The source of the CloudWatch Event.

observation_codeDeployState :: Lens' Observation (Maybe Text)

The status of the CodeDeploy deployment, for example SUCCESS or FAILURE.

observation_xRayErrorPercent :: Lens' Observation (Maybe Int)

The X-Ray request error percentage for this node.

observation_statesArn :: Lens' Observation (Maybe Text)

The Amazon Resource Name (ARN) of the step function-based observation.

observation_cloudWatchEventId :: Lens' Observation (Maybe Text)

The ID of the CloudWatch Event-based observation related to the detected problem.

observation_logText :: Lens' Observation (Maybe Text)

The log text of the observation.

observation_logFilter :: Lens' Observation (Maybe LogFilter)

The log filter of the observation.

observation_metricNamespace :: Lens' Observation (Maybe Text)

The namespace of the observation metric.

observation_rdsEventCategories :: Lens' Observation (Maybe Text)

The category of an RDS event.

observation_xRayNodeType :: Lens' Observation (Maybe Text)

The type of the X-Ray node.

observation_endTime :: Lens' Observation (Maybe UTCTime)

The time when the observation ended, in epoch seconds.

observation_statesInput :: Lens' Observation (Maybe Text)

The input to the step function-based observation.

observation_xRayNodeName :: Lens' Observation (Maybe Text)

The name of the X-Ray node.

observation_id :: Lens' Observation (Maybe Text)

The ID of the observation type.

observation_healthEventArn :: Lens' Observation (Maybe Text)

The Amazon Resource Name (ARN) of the AWS Health Event-based observation.

observation_healthEventTypeCode :: Lens' Observation (Maybe Text)

The type of the AWS Health event, for example, AWS_EC2_POWER_CONNECTIVITY_ISSUE.

observation_ebsResult :: Lens' Observation (Maybe Text)

The result of an EBS CloudWatch event, such as failed or succeeded.

observation_cloudWatchEventDetailType :: Lens' Observation (Maybe Text)

The detail type of the CloudWatch Event-based observation, for example, EC2 Instance State-change Notification.

observation_codeDeployInstanceGroupId :: Lens' Observation (Maybe Text)

The instance group to which the CodeDeploy instance belongs.

observation_ebsCause :: Lens' Observation (Maybe Text)

The cause of an EBS CloudWatch event.

observation_ebsEvent :: Lens' Observation (Maybe Text)

The type of EBS CloudWatch event, such as createVolume, deleteVolume or attachVolume.

observation_ebsRequestId :: Lens' Observation (Maybe Text)

The request ID of an EBS CloudWatch event.

observation_xRayFaultPercent :: Lens' Observation (Maybe Int)

The X-Ray request fault percentage for this node.

observation_statesExecutionArn :: Lens' Observation (Maybe Text)

The Amazon Resource Name (ARN) of the step function execution-based observation.

observation_lineTime :: Lens' Observation (Maybe UTCTime)

The timestamp in the CloudWatch Logs that specifies when the matched line occurred.

observation_unit :: Lens' Observation (Maybe Text)

The unit of the source observation metric.

observation_xRayThrottlePercent :: Lens' Observation (Maybe Int)

The X-Ray request throttle percentage for this node.

observation_healthService :: Lens' Observation (Maybe Text)

The service to which the AWS Health Event belongs, such as EC2.