Network.AWS.Athena.CreateDataCatalog

Description

Creates (registers) a data catalog with the specified name and properties. Catalogs created are visible to all users of the same Amazon Web Services account.

Creating a Request

data CreateDataCatalog

See: newCreateDataCatalog smart constructor.

Constructors

CreateDataCatalog'

Fields

tags :: Maybe [Tag]

A list of comma separated tags to add to the data catalog that is created.

description :: Maybe Text

A description of the data catalog to be created.

parameters :: Maybe (HashMap Text Text)

Specifies the Lambda function or functions to use for creating the data catalog. This is a mapping whose values depend on the catalog type.

For the HIVE data catalog type, use the following syntax. The metadata-function parameter is required. The sdk-version parameter is optional and defaults to the currently supported version.

metadata-function=lambda_arn, sdk-version=version_number

For the LAMBDA data catalog type, use one of the following sets of required parameters, but not both.

If you have one Lambda function that processes metadata and another for reading the actual data, use the following syntax. Both parameters are required.

metadata-function=lambda_arn, record-function=lambda_arn

If you have a composite Lambda function that processes both metadata and data, use the following syntax to specify your Lambda function.

function=lambda_arn

The GLUE type takes a catalog ID parameter and is required. The catalog_id is the account ID of the Amazon Web Services account to which the Glue Data Catalog belongs.

catalog-id=catalog_id

The GLUE data catalog type also applies to the default AwsDataCatalog that already exists in your account, of which you can have only one and cannot modify.

Queries that specify a Glue Data Catalog other than the default AwsDataCatalog must be run on Athena engine version 2.

In Regions where Athena engine version 2 is not available, creating new Glue data catalogs results in an INVALID_INPUT error.

name :: Text

The name of the data catalog to create. The catalog name must be unique for the Amazon Web Services account and can use a maximum of 128 alphanumeric, underscore, at sign, or hyphen characters.

type' :: DataCatalogType

The type of data catalog to create: LAMBDA for a federated catalog, HIVE for an external hive metastore, or GLUE for an Glue Data Catalog.

newCreateDataCatalog

Arguments

:: Text
$sel:name:CreateDataCatalog'

-> DataCatalogType
$sel:type':CreateDataCatalog'

-> CreateDataCatalog

Create a value of CreateDataCatalog with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateDataCatalog', createDataCatalog_tags - A list of comma separated tags to add to the data catalog that is created.

$sel:description:CreateDataCatalog', createDataCatalog_description - A description of the data catalog to be created.

$sel:parameters:CreateDataCatalog', createDataCatalog_parameters - Specifies the Lambda function or functions to use for creating the data catalog. This is a mapping whose values depend on the catalog type.

For the HIVE data catalog type, use the following syntax. The metadata-function parameter is required. The sdk-version parameter is optional and defaults to the currently supported version.

metadata-function=lambda_arn, sdk-version=version_number

For the LAMBDA data catalog type, use one of the following sets of required parameters, but not both.

If you have one Lambda function that processes metadata and another for reading the actual data, use the following syntax. Both parameters are required.

metadata-function=lambda_arn, record-function=lambda_arn

If you have a composite Lambda function that processes both metadata and data, use the following syntax to specify your Lambda function.

function=lambda_arn

The GLUE type takes a catalog ID parameter and is required. The catalog_id is the account ID of the Amazon Web Services account to which the Glue Data Catalog belongs.

catalog-id=catalog_id

The GLUE data catalog type also applies to the default AwsDataCatalog that already exists in your account, of which you can have only one and cannot modify.

Queries that specify a Glue Data Catalog other than the default AwsDataCatalog must be run on Athena engine version 2.

In Regions where Athena engine version 2 is not available, creating new Glue data catalogs results in an INVALID_INPUT error.

$sel:name:CreateDataCatalog', createDataCatalog_name - The name of the data catalog to create. The catalog name must be unique for the Amazon Web Services account and can use a maximum of 128 alphanumeric, underscore, at sign, or hyphen characters.

$sel:type':CreateDataCatalog', createDataCatalog_type - The type of data catalog to create: LAMBDA for a federated catalog, HIVE for an external hive metastore, or GLUE for an Glue Data Catalog.

Request Lenses

createDataCatalog_tags :: Lens' CreateDataCatalog (Maybe [Tag])

A list of comma separated tags to add to the data catalog that is created.

createDataCatalog_description :: Lens' CreateDataCatalog (Maybe Text)

A description of the data catalog to be created.

createDataCatalog_parameters :: Lens' CreateDataCatalog (Maybe (HashMap Text Text))

Specifies the Lambda function or functions to use for creating the data catalog. This is a mapping whose values depend on the catalog type.

For the HIVE data catalog type, use the following syntax. The metadata-function parameter is required. The sdk-version parameter is optional and defaults to the currently supported version.

metadata-function=lambda_arn, sdk-version=version_number

For the LAMBDA data catalog type, use one of the following sets of required parameters, but not both.

If you have one Lambda function that processes metadata and another for reading the actual data, use the following syntax. Both parameters are required.

metadata-function=lambda_arn, record-function=lambda_arn

If you have a composite Lambda function that processes both metadata and data, use the following syntax to specify your Lambda function.

function=lambda_arn

The GLUE type takes a catalog ID parameter and is required. The catalog_id is the account ID of the Amazon Web Services account to which the Glue Data Catalog belongs.

catalog-id=catalog_id

The GLUE data catalog type also applies to the default AwsDataCatalog that already exists in your account, of which you can have only one and cannot modify.

Queries that specify a Glue Data Catalog other than the default AwsDataCatalog must be run on Athena engine version 2.

In Regions where Athena engine version 2 is not available, creating new Glue data catalogs results in an INVALID_INPUT error.

createDataCatalog_name :: Lens' CreateDataCatalog Text

The name of the data catalog to create. The catalog name must be unique for the Amazon Web Services account and can use a maximum of 128 alphanumeric, underscore, at sign, or hyphen characters.

createDataCatalog_type :: Lens' CreateDataCatalog DataCatalogType

The type of data catalog to create: LAMBDA for a federated catalog, HIVE for an external hive metastore, or GLUE for an Glue Data Catalog.

Destructuring the Response

data CreateDataCatalogResponse

See: newCreateDataCatalogResponse smart constructor.

Constructors

CreateDataCatalogResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateDataCatalogResponse

Arguments

:: Int
$sel:httpStatus:CreateDataCatalogResponse'

-> CreateDataCatalogResponse

Create a value of CreateDataCatalogResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateDataCatalogResponse', createDataCatalogResponse_httpStatus - The response's http status code.

Response Lenses

createDataCatalogResponse_httpStatus :: Lens' CreateDataCatalogResponse Int

The response's http status code.