Network.AWS.Athena.CreateNamedQuery

Description

Creates a named query in the specified workgroup. Requires that you have access to the workgroup.

For code samples using the Amazon Web Services SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

Creating a Request

data CreateNamedQuery

See: newCreateNamedQuery smart constructor.

Constructors

CreateNamedQuery'

Fields

description :: Maybe Text

The query description.

workGroup :: Maybe Text

The name of the workgroup in which the named query is being created.

clientRequestToken :: Maybe Text

A unique case-sensitive string used to ensure the request to create the query is idempotent (executes only once). If another CreateNamedQuery request is received, the same response is returned and another query is not created. If a parameter has changed, for example, the QueryString, an error is returned.

This token is listed as not required because Amazon Web Services SDKs (for example the Amazon Web Services SDK for Java) auto-generate the token for users. If you are not using the Amazon Web Services SDK or the Amazon Web Services CLI, you must provide this token or the action will fail.

name :: Text

The query name.

database :: Text

The database to which the query belongs.

queryString :: Text

The contents of the query with all query statements.

newCreateNamedQuery

Arguments

:: Text
$sel:name:CreateNamedQuery'

-> Text
$sel:database:CreateNamedQuery'

-> Text
$sel:queryString:CreateNamedQuery'

-> CreateNamedQuery

Create a value of CreateNamedQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateNamedQuery', createNamedQuery_description - The query description.

$sel:workGroup:CreateNamedQuery', createNamedQuery_workGroup - The name of the workgroup in which the named query is being created.

$sel:clientRequestToken:CreateNamedQuery', createNamedQuery_clientRequestToken - A unique case-sensitive string used to ensure the request to create the query is idempotent (executes only once). If another CreateNamedQuery request is received, the same response is returned and another query is not created. If a parameter has changed, for example, the QueryString, an error is returned.

This token is listed as not required because Amazon Web Services SDKs (for example the Amazon Web Services SDK for Java) auto-generate the token for users. If you are not using the Amazon Web Services SDK or the Amazon Web Services CLI, you must provide this token or the action will fail.

$sel:name:CreateNamedQuery', createNamedQuery_name - The query name.

$sel:database:CreateNamedQuery', createNamedQuery_database - The database to which the query belongs.

$sel:queryString:CreateNamedQuery', createNamedQuery_queryString - The contents of the query with all query statements.

Request Lenses

createNamedQuery_description :: Lens' CreateNamedQuery (Maybe Text)

The query description.

createNamedQuery_workGroup :: Lens' CreateNamedQuery (Maybe Text)

The name of the workgroup in which the named query is being created.

createNamedQuery_clientRequestToken :: Lens' CreateNamedQuery (Maybe Text)

A unique case-sensitive string used to ensure the request to create the query is idempotent (executes only once). If another CreateNamedQuery request is received, the same response is returned and another query is not created. If a parameter has changed, for example, the QueryString, an error is returned.

This token is listed as not required because Amazon Web Services SDKs (for example the Amazon Web Services SDK for Java) auto-generate the token for users. If you are not using the Amazon Web Services SDK or the Amazon Web Services CLI, you must provide this token or the action will fail.

createNamedQuery_name :: Lens' CreateNamedQuery Text

The query name.

createNamedQuery_database :: Lens' CreateNamedQuery Text

The database to which the query belongs.

createNamedQuery_queryString :: Lens' CreateNamedQuery Text

The contents of the query with all query statements.

Destructuring the Response

data CreateNamedQueryResponse

See: newCreateNamedQueryResponse smart constructor.

Constructors

CreateNamedQueryResponse'

Fields

namedQueryId :: Maybe Text

The unique ID of the query.

httpStatus :: Int

The response's http status code.

newCreateNamedQueryResponse

Arguments

:: Int
$sel:httpStatus:CreateNamedQueryResponse'

-> CreateNamedQueryResponse

Create a value of CreateNamedQueryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namedQueryId:CreateNamedQueryResponse', createNamedQueryResponse_namedQueryId - The unique ID of the query.

$sel:httpStatus:CreateNamedQueryResponse', createNamedQueryResponse_httpStatus - The response's http status code.

Response Lenses

createNamedQueryResponse_namedQueryId :: Lens' CreateNamedQueryResponse (Maybe Text)

The unique ID of the query.

createNamedQueryResponse_httpStatus :: Lens' CreateNamedQueryResponse Int

The response's http status code.