Network.AWS.Athena.CreateWorkGroup

Description

Creates a workgroup with the specified name.

Creating a Request

data CreateWorkGroup

See: newCreateWorkGroup smart constructor.

Constructors

CreateWorkGroup'

Fields

configuration :: Maybe WorkGroupConfiguration

The configuration for the workgroup, which includes the location in Amazon S3 where query results are stored, the encryption configuration, if any, used for encrypting query results, whether the Amazon CloudWatch Metrics are enabled for the workgroup, the limit for the amount of bytes scanned (cutoff) per query, if it is specified, and whether workgroup's settings (specified with EnforceWorkGroupConfiguration) in the WorkGroupConfiguration override client-side settings. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

description :: Maybe Text

The workgroup description.

tags :: Maybe [Tag]

A list of comma separated tags to add to the workgroup that is created.

name :: Text

The workgroup name.

newCreateWorkGroup

Arguments

:: Text
$sel:name:CreateWorkGroup'

-> CreateWorkGroup

Create a value of CreateWorkGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:CreateWorkGroup', createWorkGroup_configuration - The configuration for the workgroup, which includes the location in Amazon S3 where query results are stored, the encryption configuration, if any, used for encrypting query results, whether the Amazon CloudWatch Metrics are enabled for the workgroup, the limit for the amount of bytes scanned (cutoff) per query, if it is specified, and whether workgroup's settings (specified with EnforceWorkGroupConfiguration) in the WorkGroupConfiguration override client-side settings. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

$sel:description:CreateWorkGroup', createWorkGroup_description - The workgroup description.

$sel:tags:CreateWorkGroup', createWorkGroup_tags - A list of comma separated tags to add to the workgroup that is created.

$sel:name:CreateWorkGroup', createWorkGroup_name - The workgroup name.

Request Lenses

createWorkGroup_configuration :: Lens' CreateWorkGroup (Maybe WorkGroupConfiguration)

The configuration for the workgroup, which includes the location in Amazon S3 where query results are stored, the encryption configuration, if any, used for encrypting query results, whether the Amazon CloudWatch Metrics are enabled for the workgroup, the limit for the amount of bytes scanned (cutoff) per query, if it is specified, and whether workgroup's settings (specified with EnforceWorkGroupConfiguration) in the WorkGroupConfiguration override client-side settings. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

createWorkGroup_description :: Lens' CreateWorkGroup (Maybe Text)

The workgroup description.

createWorkGroup_tags :: Lens' CreateWorkGroup (Maybe [Tag])

A list of comma separated tags to add to the workgroup that is created.

createWorkGroup_name :: Lens' CreateWorkGroup Text

The workgroup name.

Destructuring the Response

data CreateWorkGroupResponse

See: newCreateWorkGroupResponse smart constructor.

Constructors

CreateWorkGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateWorkGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateWorkGroupResponse'

-> CreateWorkGroupResponse

Create a value of CreateWorkGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateWorkGroupResponse', createWorkGroupResponse_httpStatus - The response's http status code.

Response Lenses

createWorkGroupResponse_httpStatus :: Lens' CreateWorkGroupResponse Int

The response's http status code.