Network.AWS.Athena.GetQueryResults

Description

Streams the results of a single query execution specified by QueryExecutionId from the Athena query results location in Amazon S3. For more information, see Query Results in the Amazon Athena User Guide. This request does not execute the query but returns results. Use StartQueryExecution to run a query.

To stream query results successfully, the IAM principal with permission to call GetQueryResults also must have permissions to the Amazon S3 GetObject action for the Athena query results location.

IAM principals with permission to the Amazon S3 GetObject action for the query results location are able to retrieve query results from Amazon S3 even if permission to the GetQueryResults action is denied. To restrict user or role access, ensure that Amazon S3 permissions to the Athena query location are denied.

This operation returns paginated results.

Creating a Request

data GetQueryResults

See: newGetQueryResults smart constructor.

Constructors

GetQueryResults'

Fields

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

maxResults :: Maybe Natural

The maximum number of results (rows) to return in this request.

queryExecutionId :: Text

The unique ID of the query execution.

newGetQueryResults

Arguments

:: Text
$sel:queryExecutionId:GetQueryResults'

-> GetQueryResults

Create a value of GetQueryResults with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetQueryResults', getQueryResults_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:maxResults:GetQueryResults', getQueryResults_maxResults - The maximum number of results (rows) to return in this request.

$sel:queryExecutionId:GetQueryResults', getQueryResults_queryExecutionId - The unique ID of the query execution.

Request Lenses

getQueryResults_nextToken :: Lens' GetQueryResults (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

getQueryResults_maxResults :: Lens' GetQueryResults (Maybe Natural)

The maximum number of results (rows) to return in this request.

getQueryResults_queryExecutionId :: Lens' GetQueryResults Text

The unique ID of the query execution.

Destructuring the Response

data GetQueryResultsResponse

See: newGetQueryResultsResponse smart constructor.

Constructors

GetQueryResultsResponse'

Fields

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

resultSet :: Maybe ResultSet

The results of the query execution.

updateCount :: Maybe Integer

The number of rows inserted with a CREATE TABLE AS SELECT statement.

httpStatus :: Int

The response's http status code.

newGetQueryResultsResponse

Arguments

:: Int
$sel:httpStatus:GetQueryResultsResponse'

-> GetQueryResultsResponse

Create a value of GetQueryResultsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetQueryResults', getQueryResultsResponse_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:resultSet:GetQueryResultsResponse', getQueryResultsResponse_resultSet - The results of the query execution.

$sel:updateCount:GetQueryResultsResponse', getQueryResultsResponse_updateCount - The number of rows inserted with a CREATE TABLE AS SELECT statement.

$sel:httpStatus:GetQueryResultsResponse', getQueryResultsResponse_httpStatus - The response's http status code.

Response Lenses

getQueryResultsResponse_nextToken :: Lens' GetQueryResultsResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

getQueryResultsResponse_resultSet :: Lens' GetQueryResultsResponse (Maybe ResultSet)

The results of the query execution.

getQueryResultsResponse_updateCount :: Lens' GetQueryResultsResponse (Maybe Integer)

The number of rows inserted with a CREATE TABLE AS SELECT statement.

getQueryResultsResponse_httpStatus :: Lens' GetQueryResultsResponse Int

The response's http status code.