Network.AWS.Athena.Lens

Description

Operations

CreateDataCatalog

createDataCatalog_tags :: Lens' CreateDataCatalog (Maybe [Tag])

A list of comma separated tags to add to the data catalog that is created.

createDataCatalog_description :: Lens' CreateDataCatalog (Maybe Text)

A description of the data catalog to be created.

createDataCatalog_parameters :: Lens' CreateDataCatalog (Maybe (HashMap Text Text))

Specifies the Lambda function or functions to use for creating the data catalog. This is a mapping whose values depend on the catalog type.

For the HIVE data catalog type, use the following syntax. The metadata-function parameter is required. The sdk-version parameter is optional and defaults to the currently supported version.

metadata-function=lambda_arn, sdk-version=version_number

For the LAMBDA data catalog type, use one of the following sets of required parameters, but not both.

If you have one Lambda function that processes metadata and another for reading the actual data, use the following syntax. Both parameters are required.

metadata-function=lambda_arn, record-function=lambda_arn

If you have a composite Lambda function that processes both metadata and data, use the following syntax to specify your Lambda function.

function=lambda_arn

The GLUE type takes a catalog ID parameter and is required. The catalog_id is the account ID of the Amazon Web Services account to which the Glue Data Catalog belongs.

catalog-id=catalog_id

The GLUE data catalog type also applies to the default AwsDataCatalog that already exists in your account, of which you can have only one and cannot modify.

Queries that specify a Glue Data Catalog other than the default AwsDataCatalog must be run on Athena engine version 2.

In Regions where Athena engine version 2 is not available, creating new Glue data catalogs results in an INVALID_INPUT error.

createDataCatalog_name :: Lens' CreateDataCatalog Text

The name of the data catalog to create. The catalog name must be unique for the Amazon Web Services account and can use a maximum of 128 alphanumeric, underscore, at sign, or hyphen characters.

createDataCatalog_type :: Lens' CreateDataCatalog DataCatalogType

The type of data catalog to create: LAMBDA for a federated catalog, HIVE for an external hive metastore, or GLUE for an Glue Data Catalog.

createDataCatalogResponse_httpStatus :: Lens' CreateDataCatalogResponse Int

The response's http status code.

ListQueryExecutions

listQueryExecutions_nextToken :: Lens' ListQueryExecutions (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listQueryExecutions_maxResults :: Lens' ListQueryExecutions (Maybe Natural)

The maximum number of query executions to return in this request.

listQueryExecutions_workGroup :: Lens' ListQueryExecutions (Maybe Text)

The name of the workgroup from which queries are being returned. If a workgroup is not specified, a list of available query execution IDs for the queries in the primary workgroup is returned.

listQueryExecutionsResponse_nextToken :: Lens' ListQueryExecutionsResponse (Maybe Text)

A token to be used by the next request if this request is truncated.

listQueryExecutionsResponse_queryExecutionIds :: Lens' ListQueryExecutionsResponse (Maybe (NonEmpty Text))

The unique IDs of each query execution as an array of strings.

listQueryExecutionsResponse_httpStatus :: Lens' ListQueryExecutionsResponse Int

The response's http status code.

DeletePreparedStatement

deletePreparedStatement_statementName :: Lens' DeletePreparedStatement Text

The name of the prepared statement to delete.

deletePreparedStatement_workGroup :: Lens' DeletePreparedStatement Text

The workgroup to which the statement to be deleted belongs.

deletePreparedStatementResponse_httpStatus :: Lens' DeletePreparedStatementResponse Int

The response's http status code.

ListTableMetadata

listTableMetadata_nextToken :: Lens' ListTableMetadata (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listTableMetadata_maxResults :: Lens' ListTableMetadata (Maybe Natural)

Specifies the maximum number of results to return.

listTableMetadata_expression :: Lens' ListTableMetadata (Maybe Text)

A regex filter that pattern-matches table names. If no expression is supplied, metadata for all tables are listed.

listTableMetadata_catalogName :: Lens' ListTableMetadata Text

The name of the data catalog for which table metadata should be returned.

listTableMetadata_databaseName :: Lens' ListTableMetadata Text

The name of the database for which table metadata should be returned.

listTableMetadataResponse_nextToken :: Lens' ListTableMetadataResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listTableMetadataResponse_tableMetadataList :: Lens' ListTableMetadataResponse (Maybe [TableMetadata])

A list of table metadata.

listTableMetadataResponse_httpStatus :: Lens' ListTableMetadataResponse Int

The response's http status code.

UpdatePreparedStatement

updatePreparedStatement_description :: Lens' UpdatePreparedStatement (Maybe Text)

The description of the prepared statement.

updatePreparedStatement_statementName :: Lens' UpdatePreparedStatement Text

The name of the prepared statement.

updatePreparedStatement_workGroup :: Lens' UpdatePreparedStatement Text

The workgroup for the prepared statement.

updatePreparedStatement_queryStatement :: Lens' UpdatePreparedStatement Text

The query string for the prepared statement.

updatePreparedStatementResponse_httpStatus :: Lens' UpdatePreparedStatementResponse Int

The response's http status code.

ListPreparedStatements

listPreparedStatements_nextToken :: Lens' ListPreparedStatements (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listPreparedStatements_maxResults :: Lens' ListPreparedStatements (Maybe Natural)

The maximum number of results to return in this request.

listPreparedStatements_workGroup :: Lens' ListPreparedStatements Text

The workgroup to list the prepared statements for.

listPreparedStatementsResponse_nextToken :: Lens' ListPreparedStatementsResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listPreparedStatementsResponse_preparedStatements :: Lens' ListPreparedStatementsResponse (Maybe [PreparedStatementSummary])

The list of prepared statements for the workgroup.

listPreparedStatementsResponse_httpStatus :: Lens' ListPreparedStatementsResponse Int

The response's http status code.

GetPreparedStatement

getPreparedStatement_statementName :: Lens' GetPreparedStatement Text

The name of the prepared statement to retrieve.

getPreparedStatement_workGroup :: Lens' GetPreparedStatement Text

The workgroup to which the statement to be retrieved belongs.

getPreparedStatementResponse_preparedStatement :: Lens' GetPreparedStatementResponse (Maybe PreparedStatement)

The name of the prepared statement that was retrieved.

getPreparedStatementResponse_httpStatus :: Lens' GetPreparedStatementResponse Int

The response's http status code.

GetQueryExecution

getQueryExecution_queryExecutionId :: Lens' GetQueryExecution Text

The unique ID of the query execution.

getQueryExecutionResponse_queryExecution :: Lens' GetQueryExecutionResponse (Maybe QueryExecution)

Information about the query execution.

getQueryExecutionResponse_httpStatus :: Lens' GetQueryExecutionResponse Int

The response's http status code.

GetDatabase

getDatabase_catalogName :: Lens' GetDatabase Text

The name of the data catalog that contains the database to return.

getDatabase_databaseName :: Lens' GetDatabase Text

The name of the database to return.

getDatabaseResponse_database :: Lens' GetDatabaseResponse (Maybe Database)

The database returned.

getDatabaseResponse_httpStatus :: Lens' GetDatabaseResponse Int

The response's http status code.

BatchGetNamedQuery

batchGetNamedQuery_namedQueryIds :: Lens' BatchGetNamedQuery (NonEmpty Text)

An array of query IDs.

batchGetNamedQueryResponse_namedQueries :: Lens' BatchGetNamedQueryResponse (Maybe [NamedQuery])

Information about the named query IDs submitted.

batchGetNamedQueryResponse_unprocessedNamedQueryIds :: Lens' BatchGetNamedQueryResponse (Maybe [UnprocessedNamedQueryId])

Information about provided query IDs.

batchGetNamedQueryResponse_httpStatus :: Lens' BatchGetNamedQueryResponse Int

The response's http status code.

UntagResource

untagResource_resourceARN :: Lens' UntagResource Text

Specifies the ARN of the resource from which tags are to be removed.

untagResource_tagKeys :: Lens' UntagResource [Text]

A comma-separated list of one or more tag keys whose tags are to be removed from the specified resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

TagResource

tagResource_resourceARN :: Lens' TagResource Text

Specifies the ARN of the Athena resource (workgroup or data catalog) to which tags are to be added.

tagResource_tags :: Lens' TagResource [Tag]

A collection of one or more tags, separated by commas, to be added to an Athena workgroup or data catalog resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

DeleteNamedQuery

deleteNamedQuery_namedQueryId :: Lens' DeleteNamedQuery Text

The unique ID of the query to delete.

deleteNamedQueryResponse_httpStatus :: Lens' DeleteNamedQueryResponse Int

The response's http status code.

GetDataCatalog

getDataCatalog_name :: Lens' GetDataCatalog Text

The name of the data catalog to return.

getDataCatalogResponse_dataCatalog :: Lens' GetDataCatalogResponse (Maybe DataCatalog)

The data catalog returned.

getDataCatalogResponse_httpStatus :: Lens' GetDataCatalogResponse Int

The response's http status code.

ListEngineVersions

listEngineVersions_nextToken :: Lens' ListEngineVersions (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listEngineVersions_maxResults :: Lens' ListEngineVersions (Maybe Natural)

The maximum number of engine versions to return in this request.

listEngineVersionsResponse_nextToken :: Lens' ListEngineVersionsResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listEngineVersionsResponse_engineVersions :: Lens' ListEngineVersionsResponse (Maybe [EngineVersion])

A list of engine versions that are available to choose from.

listEngineVersionsResponse_httpStatus :: Lens' ListEngineVersionsResponse Int

The response's http status code.

CreateWorkGroup

createWorkGroup_configuration :: Lens' CreateWorkGroup (Maybe WorkGroupConfiguration)

The configuration for the workgroup, which includes the location in Amazon S3 where query results are stored, the encryption configuration, if any, used for encrypting query results, whether the Amazon CloudWatch Metrics are enabled for the workgroup, the limit for the amount of bytes scanned (cutoff) per query, if it is specified, and whether workgroup's settings (specified with EnforceWorkGroupConfiguration) in the WorkGroupConfiguration override client-side settings. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

createWorkGroup_tags :: Lens' CreateWorkGroup (Maybe [Tag])

A list of comma separated tags to add to the workgroup that is created.

createWorkGroup_description :: Lens' CreateWorkGroup (Maybe Text)

The workgroup description.

createWorkGroup_name :: Lens' CreateWorkGroup Text

The workgroup name.

createWorkGroupResponse_httpStatus :: Lens' CreateWorkGroupResponse Int

The response's http status code.

ListDataCatalogs

listDataCatalogs_nextToken :: Lens' ListDataCatalogs (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listDataCatalogs_maxResults :: Lens' ListDataCatalogs (Maybe Natural)

Specifies the maximum number of data catalogs to return.

listDataCatalogsResponse_nextToken :: Lens' ListDataCatalogsResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listDataCatalogsResponse_dataCatalogsSummary :: Lens' ListDataCatalogsResponse (Maybe [DataCatalogSummary])

A summary list of data catalogs.

listDataCatalogsResponse_httpStatus :: Lens' ListDataCatalogsResponse Int

The response's http status code.

UpdateWorkGroup

updateWorkGroup_configurationUpdates :: Lens' UpdateWorkGroup (Maybe WorkGroupConfigurationUpdates)

The workgroup configuration that will be updated for the given workgroup.

updateWorkGroup_state :: Lens' UpdateWorkGroup (Maybe WorkGroupState)

The workgroup state that will be updated for the given workgroup.

updateWorkGroup_description :: Lens' UpdateWorkGroup (Maybe Text)

The workgroup description.

updateWorkGroup_workGroup :: Lens' UpdateWorkGroup Text

The specified workgroup that will be updated.

updateWorkGroupResponse_httpStatus :: Lens' UpdateWorkGroupResponse Int

The response's http status code.

GetNamedQuery

getNamedQuery_namedQueryId :: Lens' GetNamedQuery Text

The unique ID of the query. Use ListNamedQueries to get query IDs.

getNamedQueryResponse_namedQuery :: Lens' GetNamedQueryResponse (Maybe NamedQuery)

Information about the query.

getNamedQueryResponse_httpStatus :: Lens' GetNamedQueryResponse Int

The response's http status code.

ListWorkGroups

listWorkGroups_nextToken :: Lens' ListWorkGroups (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listWorkGroups_maxResults :: Lens' ListWorkGroups (Maybe Natural)

The maximum number of workgroups to return in this request.

listWorkGroupsResponse_nextToken :: Lens' ListWorkGroupsResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listWorkGroupsResponse_workGroups :: Lens' ListWorkGroupsResponse (Maybe [WorkGroupSummary])

A list of WorkGroupSummary objects that include the names, descriptions, creation times, and states for each workgroup.

listWorkGroupsResponse_httpStatus :: Lens' ListWorkGroupsResponse Int

The response's http status code.

DeleteWorkGroup

deleteWorkGroup_recursiveDeleteOption :: Lens' DeleteWorkGroup (Maybe Bool)

The option to delete the workgroup and its contents even if the workgroup contains any named queries or query executions.

deleteWorkGroup_workGroup :: Lens' DeleteWorkGroup Text

The unique name of the workgroup to delete.

deleteWorkGroupResponse_httpStatus :: Lens' DeleteWorkGroupResponse Int

The response's http status code.

ListDatabases

listDatabases_nextToken :: Lens' ListDatabases (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listDatabases_maxResults :: Lens' ListDatabases (Maybe Natural)

Specifies the maximum number of results to return.

listDatabases_catalogName :: Lens' ListDatabases Text

The name of the data catalog that contains the databases to return.

listDatabasesResponse_databaseList :: Lens' ListDatabasesResponse (Maybe [Database])

A list of databases from a data catalog.

listDatabasesResponse_nextToken :: Lens' ListDatabasesResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listDatabasesResponse_httpStatus :: Lens' ListDatabasesResponse Int

The response's http status code.

CreatePreparedStatement

createPreparedStatement_description :: Lens' CreatePreparedStatement (Maybe Text)

The description of the prepared statement.

createPreparedStatement_statementName :: Lens' CreatePreparedStatement Text

The name of the prepared statement.

createPreparedStatement_workGroup :: Lens' CreatePreparedStatement Text

The name of the workgroup to which the prepared statement belongs.

createPreparedStatement_queryStatement :: Lens' CreatePreparedStatement Text

The query string for the prepared statement.

createPreparedStatementResponse_httpStatus :: Lens' CreatePreparedStatementResponse Int

The response's http status code.

GetQueryResults

getQueryResults_nextToken :: Lens' GetQueryResults (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

getQueryResults_maxResults :: Lens' GetQueryResults (Maybe Natural)

The maximum number of results (rows) to return in this request.

getQueryResults_queryExecutionId :: Lens' GetQueryResults Text

The unique ID of the query execution.

getQueryResultsResponse_nextToken :: Lens' GetQueryResultsResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

getQueryResultsResponse_resultSet :: Lens' GetQueryResultsResponse (Maybe ResultSet)

The results of the query execution.

getQueryResultsResponse_updateCount :: Lens' GetQueryResultsResponse (Maybe Integer)

The number of rows inserted with a CREATE TABLE AS SELECT statement.

getQueryResultsResponse_httpStatus :: Lens' GetQueryResultsResponse Int

The response's http status code.

GetWorkGroup

getWorkGroup_workGroup :: Lens' GetWorkGroup Text

The name of the workgroup.

getWorkGroupResponse_workGroup :: Lens' GetWorkGroupResponse (Maybe WorkGroup)

Information about the workgroup.

getWorkGroupResponse_httpStatus :: Lens' GetWorkGroupResponse Int

The response's http status code.

StopQueryExecution

stopQueryExecution_queryExecutionId :: Lens' StopQueryExecution Text

The unique ID of the query execution to stop.

stopQueryExecutionResponse_httpStatus :: Lens' StopQueryExecutionResponse Int

The response's http status code.

StartQueryExecution

startQueryExecution_queryExecutionContext :: Lens' StartQueryExecution (Maybe QueryExecutionContext)

The database within which the query executes.

startQueryExecution_resultConfiguration :: Lens' StartQueryExecution (Maybe ResultConfiguration)

Specifies information about where and how to save the results of the query execution. If the query runs in a workgroup, then workgroup's settings may override query settings. This affects the query results location. The workgroup settings override is specified in EnforceWorkGroupConfiguration (true/false) in the WorkGroupConfiguration. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

startQueryExecution_workGroup :: Lens' StartQueryExecution (Maybe Text)

The name of the workgroup in which the query is being started.

startQueryExecution_clientRequestToken :: Lens' StartQueryExecution (Maybe Text)

A unique case-sensitive string used to ensure the request to create the query is idempotent (executes only once). If another StartQueryExecution request is received, the same response is returned and another query is not created. If a parameter has changed, for example, the QueryString, an error is returned.

This token is listed as not required because Amazon Web Services SDKs (for example the Amazon Web Services SDK for Java) auto-generate the token for users. If you are not using the Amazon Web Services SDK or the Amazon Web Services CLI, you must provide this token or the action will fail.

startQueryExecution_queryString :: Lens' StartQueryExecution Text

The SQL query statements to be executed.

startQueryExecutionResponse_queryExecutionId :: Lens' StartQueryExecutionResponse (Maybe Text)

The unique ID of the query that ran as a result of this request.

startQueryExecutionResponse_httpStatus :: Lens' StartQueryExecutionResponse Int

The response's http status code.

ListNamedQueries

listNamedQueries_nextToken :: Lens' ListNamedQueries (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listNamedQueries_maxResults :: Lens' ListNamedQueries (Maybe Natural)

The maximum number of queries to return in this request.

listNamedQueries_workGroup :: Lens' ListNamedQueries (Maybe Text)

The name of the workgroup from which the named queries are being returned. If a workgroup is not specified, the saved queries for the primary workgroup are returned.

listNamedQueriesResponse_namedQueryIds :: Lens' ListNamedQueriesResponse (Maybe (NonEmpty Text))

The list of unique query IDs.

listNamedQueriesResponse_nextToken :: Lens' ListNamedQueriesResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listNamedQueriesResponse_httpStatus :: Lens' ListNamedQueriesResponse Int

The response's http status code.

GetTableMetadata

getTableMetadata_catalogName :: Lens' GetTableMetadata Text

The name of the data catalog that contains the database and table metadata to return.

getTableMetadata_databaseName :: Lens' GetTableMetadata Text

The name of the database that contains the table metadata to return.

getTableMetadata_tableName :: Lens' GetTableMetadata Text

The name of the table for which metadata is returned.

getTableMetadataResponse_tableMetadata :: Lens' GetTableMetadataResponse (Maybe TableMetadata)

An object that contains table metadata.

getTableMetadataResponse_httpStatus :: Lens' GetTableMetadataResponse Int

The response's http status code.

CreateNamedQuery

createNamedQuery_description :: Lens' CreateNamedQuery (Maybe Text)

The query description.

createNamedQuery_workGroup :: Lens' CreateNamedQuery (Maybe Text)

The name of the workgroup in which the named query is being created.

createNamedQuery_clientRequestToken :: Lens' CreateNamedQuery (Maybe Text)

A unique case-sensitive string used to ensure the request to create the query is idempotent (executes only once). If another CreateNamedQuery request is received, the same response is returned and another query is not created. If a parameter has changed, for example, the QueryString, an error is returned.

This token is listed as not required because Amazon Web Services SDKs (for example the Amazon Web Services SDK for Java) auto-generate the token for users. If you are not using the Amazon Web Services SDK or the Amazon Web Services CLI, you must provide this token or the action will fail.

createNamedQuery_name :: Lens' CreateNamedQuery Text

The query name.

createNamedQuery_database :: Lens' CreateNamedQuery Text

The database to which the query belongs.

createNamedQuery_queryString :: Lens' CreateNamedQuery Text

The contents of the query with all query statements.

createNamedQueryResponse_namedQueryId :: Lens' CreateNamedQueryResponse (Maybe Text)

The unique ID of the query.

createNamedQueryResponse_httpStatus :: Lens' CreateNamedQueryResponse Int

The response's http status code.

BatchGetQueryExecution

batchGetQueryExecution_queryExecutionIds :: Lens' BatchGetQueryExecution (NonEmpty Text)

An array of query execution IDs.

batchGetQueryExecutionResponse_queryExecutions :: Lens' BatchGetQueryExecutionResponse (Maybe [QueryExecution])

Information about a query execution.

batchGetQueryExecutionResponse_unprocessedQueryExecutionIds :: Lens' BatchGetQueryExecutionResponse (Maybe [UnprocessedQueryExecutionId])

Information about the query executions that failed to run.

batchGetQueryExecutionResponse_httpStatus :: Lens' BatchGetQueryExecutionResponse Int

The response's http status code.

DeleteDataCatalog

deleteDataCatalog_name :: Lens' DeleteDataCatalog Text

The name of the data catalog to delete.

deleteDataCatalogResponse_httpStatus :: Lens' DeleteDataCatalogResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_nextToken :: Lens' ListTagsForResource (Maybe Text)

The token for the next set of results, or null if there are no additional results for this request, where the request lists the tags for the resource with the specified ARN.

listTagsForResource_maxResults :: Lens' ListTagsForResource (Maybe Natural)

The maximum number of results to be returned per request that lists the tags for the resource.

listTagsForResource_resourceARN :: Lens' ListTagsForResource Text

Lists the tags for the resource with the specified ARN.

listTagsForResourceResponse_nextToken :: Lens' ListTagsForResourceResponse (Maybe Text)

A token to be used by the next request if this request is truncated.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

The list of tags associated with the specified resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

UpdateDataCatalog

updateDataCatalog_description :: Lens' UpdateDataCatalog (Maybe Text)

New or modified text that describes the data catalog.

updateDataCatalog_parameters :: Lens' UpdateDataCatalog (Maybe (HashMap Text Text))

Specifies the Lambda function or functions to use for updating the data catalog. This is a mapping whose values depend on the catalog type.

For the HIVE data catalog type, use the following syntax. The metadata-function parameter is required. The sdk-version parameter is optional and defaults to the currently supported version.

metadata-function=lambda_arn, sdk-version=version_number

For the LAMBDA data catalog type, use one of the following sets of required parameters, but not both.

If you have one Lambda function that processes metadata and another for reading the actual data, use the following syntax. Both parameters are required.

metadata-function=lambda_arn, record-function=lambda_arn

If you have a composite Lambda function that processes both metadata and data, use the following syntax to specify your Lambda function.

function=lambda_arn

updateDataCatalog_name :: Lens' UpdateDataCatalog Text

The name of the data catalog to update. The catalog name must be unique for the Amazon Web Services account and can use a maximum of 128 alphanumeric, underscore, at sign, or hyphen characters.

updateDataCatalog_type :: Lens' UpdateDataCatalog DataCatalogType

Specifies the type of data catalog to update. Specify LAMBDA for a federated catalog, HIVE for an external hive metastore, or GLUE for an Glue Data Catalog.

updateDataCatalogResponse_httpStatus :: Lens' UpdateDataCatalogResponse Int

The response's http status code.

Types

Column

column_comment :: Lens' Column (Maybe Text)

Optional information about the column.

column_type :: Lens' Column (Maybe Text)

The data type of the column.

column_name :: Lens' Column Text

The name of the column.

ColumnInfo

columnInfo_catalogName :: Lens' ColumnInfo (Maybe Text)

The catalog to which the query results belong.

columnInfo_tableName :: Lens' ColumnInfo (Maybe Text)

The table name for the query results.

columnInfo_precision :: Lens' ColumnInfo (Maybe Int)

For DECIMAL data types, specifies the total number of digits, up to 38. For performance reasons, we recommend up to 18 digits.

columnInfo_caseSensitive :: Lens' ColumnInfo (Maybe Bool)

Indicates whether values in the column are case-sensitive.

columnInfo_nullable :: Lens' ColumnInfo (Maybe ColumnNullable)

Indicates the column's nullable status.

columnInfo_label :: Lens' ColumnInfo (Maybe Text)

A column label.

columnInfo_schemaName :: Lens' ColumnInfo (Maybe Text)

The schema name (database name) to which the query results belong.

columnInfo_scale :: Lens' ColumnInfo (Maybe Int)

For DECIMAL data types, specifies the total number of digits in the fractional part of the value. Defaults to 0.

columnInfo_name :: Lens' ColumnInfo Text

The name of the column.

columnInfo_type :: Lens' ColumnInfo Text

The data type of the column.

DataCatalog

dataCatalog_description :: Lens' DataCatalog (Maybe Text)

An optional description of the data catalog.

dataCatalog_parameters :: Lens' DataCatalog (Maybe (HashMap Text Text))

Specifies the Lambda function or functions to use for the data catalog. This is a mapping whose values depend on the catalog type.

For the HIVE data catalog type, use the following syntax. The metadata-function parameter is required. The sdk-version parameter is optional and defaults to the currently supported version.

metadata-function=lambda_arn, sdk-version=version_number

For the LAMBDA data catalog type, use one of the following sets of required parameters, but not both.

If you have one Lambda function that processes metadata and another for reading the actual data, use the following syntax. Both parameters are required.

metadata-function=lambda_arn, record-function=lambda_arn

If you have a composite Lambda function that processes both metadata and data, use the following syntax to specify your Lambda function.

function=lambda_arn

The GLUE type takes a catalog ID parameter and is required. The catalog_id is the account ID of the Amazon Web Services account to which the Glue catalog belongs.

catalog-id=catalog_id

The GLUE data catalog type also applies to the default AwsDataCatalog that already exists in your account, of which you can have only one and cannot modify.

Queries that specify a Glue Data Catalog other than the default AwsDataCatalog must be run on Athena engine version 2.

dataCatalog_name :: Lens' DataCatalog Text

The name of the data catalog. The catalog name must be unique for the Amazon Web Services account and can use a maximum of 128 alphanumeric, underscore, at sign, or hyphen characters.

dataCatalog_type :: Lens' DataCatalog DataCatalogType

The type of data catalog to create: LAMBDA for a federated catalog, HIVE for an external hive metastore, or GLUE for an Glue Data Catalog.

DataCatalogSummary

dataCatalogSummary_catalogName :: Lens' DataCatalogSummary (Maybe Text)

The name of the data catalog.

dataCatalogSummary_type :: Lens' DataCatalogSummary (Maybe DataCatalogType)

The data catalog type.

Database

database_description :: Lens' Database (Maybe Text)

An optional description of the database.

database_parameters :: Lens' Database (Maybe (HashMap Text Text))

A set of custom key/value pairs.

database_name :: Lens' Database Text

The name of the database.

Datum

datum_varCharValue :: Lens' Datum (Maybe Text)

The value of the datum.

EncryptionConfiguration

encryptionConfiguration_kmsKey :: Lens' EncryptionConfiguration (Maybe Text)

For SSE-KMS and CSE-KMS, this is the KMS key ARN or ID.

encryptionConfiguration_encryptionOption :: Lens' EncryptionConfiguration EncryptionOption

Indicates whether Amazon S3 server-side encryption with Amazon S3-managed keys (SSE-S3), server-side encryption with KMS-managed keys (SSE-KMS), or client-side encryption with KMS-managed keys (CSE-KMS) is used.

If a query runs in a workgroup and the workgroup overrides client-side settings, then the workgroup's setting for encryption is used. It specifies whether query results must be encrypted, for all queries that run in this workgroup.

EngineVersion

engineVersion_effectiveEngineVersion :: Lens' EngineVersion (Maybe Text)

Read only. The engine version on which the query runs. If the user requests a valid engine version other than Auto, the effective engine version is the same as the engine version that the user requested. If the user requests Auto, the effective engine version is chosen by Athena. When a request to update the engine version is made by a CreateWorkGroup or UpdateWorkGroup operation, the EffectiveEngineVersion field is ignored.

engineVersion_selectedEngineVersion :: Lens' EngineVersion (Maybe Text)

The engine version requested by the user. Possible values are determined by the output of ListEngineVersions, including Auto. The default is Auto.

NamedQuery

namedQuery_namedQueryId :: Lens' NamedQuery (Maybe Text)

The unique identifier of the query.

namedQuery_description :: Lens' NamedQuery (Maybe Text)

The query description.

namedQuery_workGroup :: Lens' NamedQuery (Maybe Text)

The name of the workgroup that contains the named query.

namedQuery_name :: Lens' NamedQuery Text

The query name.

namedQuery_database :: Lens' NamedQuery Text

The database to which the query belongs.

namedQuery_queryString :: Lens' NamedQuery Text

The SQL query statements that comprise the query.

PreparedStatement

preparedStatement_workGroupName :: Lens' PreparedStatement (Maybe Text)

The name of the workgroup to which the prepared statement belongs.

preparedStatement_statementName :: Lens' PreparedStatement (Maybe Text)

The name of the prepared statement.

preparedStatement_queryStatement :: Lens' PreparedStatement (Maybe Text)

The query string for the prepared statement.

preparedStatement_lastModifiedTime :: Lens' PreparedStatement (Maybe UTCTime)

The last modified time of the prepared statement.

preparedStatement_description :: Lens' PreparedStatement (Maybe Text)

The description of the prepared statement.

PreparedStatementSummary

preparedStatementSummary_statementName :: Lens' PreparedStatementSummary (Maybe Text)

The name of the prepared statement.

preparedStatementSummary_lastModifiedTime :: Lens' PreparedStatementSummary (Maybe UTCTime)

The last modified time of the prepared statement.

QueryExecution

queryExecution_status :: Lens' QueryExecution (Maybe QueryExecutionStatus)

The completion date, current state, submission time, and state change reason (if applicable) for the query execution.

queryExecution_queryExecutionId :: Lens' QueryExecution (Maybe Text)

The unique identifier for each query execution.

queryExecution_statistics :: Lens' QueryExecution (Maybe QueryExecutionStatistics)

Query execution statistics, such as the amount of data scanned, the amount of time that the query took to process, and the type of statement that was run.

queryExecution_query :: Lens' QueryExecution (Maybe Text)

The SQL query statements which the query execution ran.

queryExecution_queryExecutionContext :: Lens' QueryExecution (Maybe QueryExecutionContext)

The database in which the query execution occurred.

queryExecution_engineVersion :: Lens' QueryExecution (Maybe EngineVersion)

The engine version that executed the query.

queryExecution_resultConfiguration :: Lens' QueryExecution (Maybe ResultConfiguration)

The location in Amazon S3 where query results were stored and the encryption option, if any, used for query results. These are known as "client-side settings". If workgroup settings override client-side settings, then the query uses the location for the query results and the encryption configuration that are specified for the workgroup.

queryExecution_workGroup :: Lens' QueryExecution (Maybe Text)

The name of the workgroup in which the query ran.

queryExecution_statementType :: Lens' QueryExecution (Maybe StatementType)

The type of query statement that was run. DDL indicates DDL query statements. DML indicates DML (Data Manipulation Language) query statements, such as CREATE TABLE AS SELECT. UTILITY indicates query statements other than DDL and DML, such as SHOW CREATE TABLE, or DESCRIBE TABLE.

QueryExecutionContext

queryExecutionContext_catalog :: Lens' QueryExecutionContext (Maybe Text)

The name of the data catalog used in the query execution.

queryExecutionContext_database :: Lens' QueryExecutionContext (Maybe Text)

The name of the database used in the query execution. The database must exist in the catalog.

QueryExecutionStatistics

queryExecutionStatistics_totalExecutionTimeInMillis :: Lens' QueryExecutionStatistics (Maybe Integer)

The number of milliseconds that Athena took to run the query.

queryExecutionStatistics_serviceProcessingTimeInMillis :: Lens' QueryExecutionStatistics (Maybe Integer)

The number of milliseconds that Athena took to finalize and publish the query results after the query engine finished running the query.

queryExecutionStatistics_queryQueueTimeInMillis :: Lens' QueryExecutionStatistics (Maybe Integer)

The number of milliseconds that the query was in your query queue waiting for resources. Note that if transient errors occur, Athena might automatically add the query back to the queue.

queryExecutionStatistics_dataScannedInBytes :: Lens' QueryExecutionStatistics (Maybe Integer)

The number of bytes in the data that was queried.

queryExecutionStatistics_queryPlanningTimeInMillis :: Lens' QueryExecutionStatistics (Maybe Integer)

The number of milliseconds that Athena took to plan the query processing flow. This includes the time spent retrieving table partitions from the data source. Note that because the query engine performs the query planning, query planning time is a subset of engine processing time.

queryExecutionStatistics_engineExecutionTimeInMillis :: Lens' QueryExecutionStatistics (Maybe Integer)

The number of milliseconds that the query took to execute.

queryExecutionStatistics_dataManifestLocation :: Lens' QueryExecutionStatistics (Maybe Text)

The location and file name of a data manifest file. The manifest file is saved to the Athena query results location in Amazon S3. The manifest file tracks files that the query wrote to Amazon S3. If the query fails, the manifest file also tracks files that the query intended to write. The manifest is useful for identifying orphaned files resulting from a failed query. For more information, see Working with Query Results, Output Files, and Query History in the Amazon Athena User Guide.

QueryExecutionStatus

queryExecutionStatus_submissionDateTime :: Lens' QueryExecutionStatus (Maybe UTCTime)

The date and time that the query was submitted.

queryExecutionStatus_completionDateTime :: Lens' QueryExecutionStatus (Maybe UTCTime)

The date and time that the query completed.

queryExecutionStatus_stateChangeReason :: Lens' QueryExecutionStatus (Maybe Text)

Further detail about the status of the query.

queryExecutionStatus_state :: Lens' QueryExecutionStatus (Maybe QueryExecutionState)

The state of query execution. QUEUED indicates that the query has been submitted to the service, and Athena will execute the query as soon as resources are available. RUNNING indicates that the query is in execution phase. SUCCEEDED indicates that the query completed without errors. FAILED indicates that the query experienced an error and did not complete processing. CANCELLED indicates that a user input interrupted query execution.

Athena automatically retries your queries in cases of certain transient errors. As a result, you may see the query state transition from RUNNING or FAILED to QUEUED.

ResultConfiguration

resultConfiguration_encryptionConfiguration :: Lens' ResultConfiguration (Maybe EncryptionConfiguration)

If query results are encrypted in Amazon S3, indicates the encryption option used (for example, SSE-KMS or CSE-KMS) and key information. This is a client-side setting. If workgroup settings override client-side settings, then the query uses the encryption configuration that is specified for the workgroup, and also uses the location for storing query results specified in the workgroup. See WorkGroupConfiguration$EnforceWorkGroupConfiguration and Workgroup Settings Override Client-Side Settings.

resultConfiguration_outputLocation :: Lens' ResultConfiguration (Maybe Text)

The location in Amazon S3 where your query results are stored, such as s3://path/to/query/bucket/. To run the query, you must specify the query results location using one of the ways: either for individual queries using either this setting (client-side), or in the workgroup, using WorkGroupConfiguration. If none of them is set, Athena issues an error that no output location is provided. For more information, see Query Results. If workgroup settings override client-side settings, then the query uses the settings specified for the workgroup. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

ResultConfigurationUpdates

resultConfigurationUpdates_encryptionConfiguration :: Lens' ResultConfigurationUpdates (Maybe EncryptionConfiguration)

The encryption configuration for the query results.

resultConfigurationUpdates_removeOutputLocation :: Lens' ResultConfigurationUpdates (Maybe Bool)

If set to "true", indicates that the previously-specified query results location (also known as a client-side setting) for queries in this workgroup should be ignored and set to null. If set to "false" or not set, and a value is present in the OutputLocation in ResultConfigurationUpdates (the client-side setting), the OutputLocation in the workgroup's ResultConfiguration will be updated with the new value. For more information, see Workgroup Settings Override Client-Side Settings.

resultConfigurationUpdates_removeEncryptionConfiguration :: Lens' ResultConfigurationUpdates (Maybe Bool)

If set to "true", indicates that the previously-specified encryption configuration (also known as the client-side setting) for queries in this workgroup should be ignored and set to null. If set to "false" or not set, and a value is present in the EncryptionConfiguration in ResultConfigurationUpdates (the client-side setting), the EncryptionConfiguration in the workgroup's ResultConfiguration will be updated with the new value. For more information, see Workgroup Settings Override Client-Side Settings.

resultConfigurationUpdates_outputLocation :: Lens' ResultConfigurationUpdates (Maybe Text)

The location in Amazon S3 where your query results are stored, such as s3://path/to/query/bucket/. For more information, see Query Results If workgroup settings override client-side settings, then the query uses the location for the query results and the encryption configuration that are specified for the workgroup. The "workgroup settings override" is specified in EnforceWorkGroupConfiguration (true/false) in the WorkGroupConfiguration. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

ResultSet

resultSet_rows :: Lens' ResultSet (Maybe [Row])

The rows in the table.

resultSet_resultSetMetadata :: Lens' ResultSet (Maybe ResultSetMetadata)

The metadata that describes the column structure and data types of a table of query results.

ResultSetMetadata

resultSetMetadata_columnInfo :: Lens' ResultSetMetadata (Maybe [ColumnInfo])

Information about the columns returned in a query result metadata.

Row

row_data :: Lens' Row (Maybe [Datum])

The data that populates a row in a query result table.

TableMetadata

tableMetadata_tableType :: Lens' TableMetadata (Maybe Text)

The type of table. In Athena, only EXTERNAL_TABLE is supported.

tableMetadata_createTime :: Lens' TableMetadata (Maybe UTCTime)

The time that the table was created.

tableMetadata_partitionKeys :: Lens' TableMetadata (Maybe [Column])

A list of the partition keys in the table.

tableMetadata_lastAccessTime :: Lens' TableMetadata (Maybe UTCTime)

The last time the table was accessed.

tableMetadata_columns :: Lens' TableMetadata (Maybe [Column])

A list of the columns in the table.

tableMetadata_parameters :: Lens' TableMetadata (Maybe (HashMap Text Text))

A set of custom key/value pairs for table properties.

tableMetadata_name :: Lens' TableMetadata Text

The name of the table.

Tag

tag_key :: Lens' Tag (Maybe Text)

A tag key. The tag key length is from 1 to 128 Unicode characters in UTF-8. You can use letters and numbers representable in UTF-8, and the following characters: + - = . _ : / @. Tag keys are case-sensitive and must be unique per resource.

tag_value :: Lens' Tag (Maybe Text)

A tag value. The tag value length is from 0 to 256 Unicode characters in UTF-8. You can use letters and numbers representable in UTF-8, and the following characters: + - = . _ : / @. Tag values are case-sensitive.

UnprocessedNamedQueryId

unprocessedNamedQueryId_namedQueryId :: Lens' UnprocessedNamedQueryId (Maybe Text)

The unique identifier of the named query.

unprocessedNamedQueryId_errorMessage :: Lens' UnprocessedNamedQueryId (Maybe Text)

The error message returned when the processing request for the named query failed, if applicable.

unprocessedNamedQueryId_errorCode :: Lens' UnprocessedNamedQueryId (Maybe Text)

The error code returned when the processing request for the named query failed, if applicable.

UnprocessedQueryExecutionId

unprocessedQueryExecutionId_queryExecutionId :: Lens' UnprocessedQueryExecutionId (Maybe Text)

The unique identifier of the query execution.

unprocessedQueryExecutionId_errorMessage :: Lens' UnprocessedQueryExecutionId (Maybe Text)

The error message returned when the query execution failed to process, if applicable.

unprocessedQueryExecutionId_errorCode :: Lens' UnprocessedQueryExecutionId (Maybe Text)

The error code returned when the query execution failed to process, if applicable.

WorkGroup

workGroup_creationTime :: Lens' WorkGroup (Maybe UTCTime)

The date and time the workgroup was created.

workGroup_configuration :: Lens' WorkGroup (Maybe WorkGroupConfiguration)

The configuration of the workgroup, which includes the location in Amazon S3 where query results are stored, the encryption configuration, if any, used for query results; whether the Amazon CloudWatch Metrics are enabled for the workgroup; whether workgroup settings override client-side settings; and the data usage limits for the amount of data scanned per query or per workgroup. The workgroup settings override is specified in EnforceWorkGroupConfiguration (true/false) in the WorkGroupConfiguration. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

workGroup_state :: Lens' WorkGroup (Maybe WorkGroupState)

The state of the workgroup: ENABLED or DISABLED.

workGroup_description :: Lens' WorkGroup (Maybe Text)

The workgroup description.

workGroup_name :: Lens' WorkGroup Text

The workgroup name.

WorkGroupConfiguration

workGroupConfiguration_bytesScannedCutoffPerQuery :: Lens' WorkGroupConfiguration (Maybe Natural)

The upper data usage limit (cutoff) for the amount of bytes a single query in a workgroup is allowed to scan.

workGroupConfiguration_publishCloudWatchMetricsEnabled :: Lens' WorkGroupConfiguration (Maybe Bool)

Indicates that the Amazon CloudWatch metrics are enabled for the workgroup.

workGroupConfiguration_requesterPaysEnabled :: Lens' WorkGroupConfiguration (Maybe Bool)

If set to true, allows members assigned to a workgroup to reference Amazon S3 Requester Pays buckets in queries. If set to false, workgroup members cannot query data from Requester Pays buckets, and queries that retrieve data from Requester Pays buckets cause an error. The default is false. For more information about Requester Pays buckets, see Requester Pays Buckets in the Amazon Simple Storage Service Developer Guide.

workGroupConfiguration_enforceWorkGroupConfiguration :: Lens' WorkGroupConfiguration (Maybe Bool)

If set to "true", the settings for the workgroup override client-side settings. If set to "false", client-side settings are used. For more information, see Workgroup Settings Override Client-Side Settings.

workGroupConfiguration_engineVersion :: Lens' WorkGroupConfiguration (Maybe EngineVersion)

The engine version that all queries running on the workgroup use. Queries on the AmazonAthenaPreviewFunctionality workgroup run on the preview engine regardless of this setting.

workGroupConfiguration_resultConfiguration :: Lens' WorkGroupConfiguration (Maybe ResultConfiguration)

The configuration for the workgroup, which includes the location in Amazon S3 where query results are stored and the encryption option, if any, used for query results. To run the query, you must specify the query results location using one of the ways: either in the workgroup using this setting, or for individual queries (client-side), using ResultConfiguration$OutputLocation. If none of them is set, Athena issues an error that no output location is provided. For more information, see Query Results.

WorkGroupConfigurationUpdates

workGroupConfigurationUpdates_bytesScannedCutoffPerQuery :: Lens' WorkGroupConfigurationUpdates (Maybe Natural)

The upper limit (cutoff) for the amount of bytes a single query in a workgroup is allowed to scan.

workGroupConfigurationUpdates_publishCloudWatchMetricsEnabled :: Lens' WorkGroupConfigurationUpdates (Maybe Bool)

Indicates whether this workgroup enables publishing metrics to Amazon CloudWatch.

workGroupConfigurationUpdates_resultConfigurationUpdates :: Lens' WorkGroupConfigurationUpdates (Maybe ResultConfigurationUpdates)

The result configuration information about the queries in this workgroup that will be updated. Includes the updated results location and an updated option for encrypting query results.

workGroupConfigurationUpdates_requesterPaysEnabled :: Lens' WorkGroupConfigurationUpdates (Maybe Bool)

If set to true, allows members assigned to a workgroup to specify Amazon S3 Requester Pays buckets in queries. If set to false, workgroup members cannot query data from Requester Pays buckets, and queries that retrieve data from Requester Pays buckets cause an error. The default is false. For more information about Requester Pays buckets, see Requester Pays Buckets in the Amazon Simple Storage Service Developer Guide.

workGroupConfigurationUpdates_enforceWorkGroupConfiguration :: Lens' WorkGroupConfigurationUpdates (Maybe Bool)

If set to "true", the settings for the workgroup override client-side settings. If set to "false" client-side settings are used. For more information, see Workgroup Settings Override Client-Side Settings.

workGroupConfigurationUpdates_removeBytesScannedCutoffPerQuery :: Lens' WorkGroupConfigurationUpdates (Maybe Bool)

Indicates that the data usage control limit per query is removed. WorkGroupConfiguration$BytesScannedCutoffPerQuery

workGroupConfigurationUpdates_engineVersion :: Lens' WorkGroupConfigurationUpdates (Maybe EngineVersion)

The engine version requested when a workgroup is updated. After the update, all queries on the workgroup run on the requested engine version. If no value was previously set, the default is Auto. Queries on the AmazonAthenaPreviewFunctionality workgroup run on the preview engine regardless of this setting.

WorkGroupSummary

workGroupSummary_creationTime :: Lens' WorkGroupSummary (Maybe UTCTime)

The workgroup creation date and time.

workGroupSummary_name :: Lens' WorkGroupSummary (Maybe Text)

The name of the workgroup.

workGroupSummary_state :: Lens' WorkGroupSummary (Maybe WorkGroupState)

The state of the workgroup.

workGroupSummary_engineVersion :: Lens' WorkGroupSummary (Maybe EngineVersion)

The engine version setting for all queries on the workgroup. Queries on the AmazonAthenaPreviewFunctionality workgroup run on the preview engine regardless of this setting.

workGroupSummary_description :: Lens' WorkGroupSummary (Maybe Text)

The workgroup description.