Network.AWS.Athena.ListNamedQueries

Description

Provides a list of available query IDs only for queries saved in the specified workgroup. Requires that you have access to the specified workgroup. If a workgroup is not specified, lists the saved queries for the primary workgroup.

For code samples using the Amazon Web Services SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

This operation returns paginated results.

Creating a Request

data ListNamedQueries

See: newListNamedQueries smart constructor.

Constructors

ListNamedQueries'

Fields

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

maxResults :: Maybe Natural

The maximum number of queries to return in this request.

workGroup :: Maybe Text

The name of the workgroup from which the named queries are being returned. If a workgroup is not specified, the saved queries for the primary workgroup are returned.

newListNamedQueries :: ListNamedQueries

Create a value of ListNamedQueries with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListNamedQueries', listNamedQueries_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:maxResults:ListNamedQueries', listNamedQueries_maxResults - The maximum number of queries to return in this request.

$sel:workGroup:ListNamedQueries', listNamedQueries_workGroup - The name of the workgroup from which the named queries are being returned. If a workgroup is not specified, the saved queries for the primary workgroup are returned.

Request Lenses

listNamedQueries_nextToken :: Lens' ListNamedQueries (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listNamedQueries_maxResults :: Lens' ListNamedQueries (Maybe Natural)

The maximum number of queries to return in this request.

listNamedQueries_workGroup :: Lens' ListNamedQueries (Maybe Text)

The name of the workgroup from which the named queries are being returned. If a workgroup is not specified, the saved queries for the primary workgroup are returned.

Destructuring the Response

data ListNamedQueriesResponse

See: newListNamedQueriesResponse smart constructor.

Constructors

ListNamedQueriesResponse'

Fields

namedQueryIds :: Maybe (NonEmpty Text)

The list of unique query IDs.

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

httpStatus :: Int

The response's http status code.

newListNamedQueriesResponse

Arguments

:: Int
$sel:httpStatus:ListNamedQueriesResponse'

-> ListNamedQueriesResponse

Create a value of ListNamedQueriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namedQueryIds:ListNamedQueriesResponse', listNamedQueriesResponse_namedQueryIds - The list of unique query IDs.

$sel:nextToken:ListNamedQueries', listNamedQueriesResponse_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:httpStatus:ListNamedQueriesResponse', listNamedQueriesResponse_httpStatus - The response's http status code.

Response Lenses

listNamedQueriesResponse_namedQueryIds :: Lens' ListNamedQueriesResponse (Maybe (NonEmpty Text))

The list of unique query IDs.

listNamedQueriesResponse_nextToken :: Lens' ListNamedQueriesResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listNamedQueriesResponse_httpStatus :: Lens' ListNamedQueriesResponse Int

The response's http status code.