Network.AWS.Athena.ListPreparedStatements

Description

Lists the prepared statements in the specfied workgroup.

Creating a Request

data ListPreparedStatements

See: newListPreparedStatements smart constructor.

Constructors

ListPreparedStatements'

Fields

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

maxResults :: Maybe Natural

The maximum number of results to return in this request.

workGroup :: Text

The workgroup to list the prepared statements for.

newListPreparedStatements

Arguments

:: Text
$sel:workGroup:ListPreparedStatements'

-> ListPreparedStatements

Create a value of ListPreparedStatements with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPreparedStatements', listPreparedStatements_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:maxResults:ListPreparedStatements', listPreparedStatements_maxResults - The maximum number of results to return in this request.

$sel:workGroup:ListPreparedStatements', listPreparedStatements_workGroup - The workgroup to list the prepared statements for.

Request Lenses

listPreparedStatements_nextToken :: Lens' ListPreparedStatements (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listPreparedStatements_maxResults :: Lens' ListPreparedStatements (Maybe Natural)

The maximum number of results to return in this request.

listPreparedStatements_workGroup :: Lens' ListPreparedStatements Text

The workgroup to list the prepared statements for.

Destructuring the Response

data ListPreparedStatementsResponse

See: newListPreparedStatementsResponse smart constructor.

Constructors

ListPreparedStatementsResponse'

Fields

preparedStatements :: Maybe [PreparedStatementSummary]

The list of prepared statements for the workgroup.

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

httpStatus :: Int

The response's http status code.

newListPreparedStatementsResponse

Arguments

:: Int
$sel:httpStatus:ListPreparedStatementsResponse'

-> ListPreparedStatementsResponse

Create a value of ListPreparedStatementsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preparedStatements:ListPreparedStatementsResponse', listPreparedStatementsResponse_preparedStatements - The list of prepared statements for the workgroup.

$sel:nextToken:ListPreparedStatements', listPreparedStatementsResponse_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:httpStatus:ListPreparedStatementsResponse', listPreparedStatementsResponse_httpStatus - The response's http status code.

Response Lenses

listPreparedStatementsResponse_preparedStatements :: Lens' ListPreparedStatementsResponse (Maybe [PreparedStatementSummary])

The list of prepared statements for the workgroup.

listPreparedStatementsResponse_nextToken :: Lens' ListPreparedStatementsResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listPreparedStatementsResponse_httpStatus :: Lens' ListPreparedStatementsResponse Int

The response's http status code.