Network.AWS.Athena.ListQueryExecutions

Description

Provides a list of available query execution IDs for the queries in the specified workgroup. If a workgroup is not specified, returns a list of query execution IDs for the primary workgroup. Requires you to have access to the workgroup in which the queries ran.

For code samples using the Amazon Web Services SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

This operation returns paginated results.

Creating a Request

data ListQueryExecutions

See: newListQueryExecutions smart constructor.

Constructors

ListQueryExecutions'

Fields

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

maxResults :: Maybe Natural

The maximum number of query executions to return in this request.

workGroup :: Maybe Text

The name of the workgroup from which queries are being returned. If a workgroup is not specified, a list of available query execution IDs for the queries in the primary workgroup is returned.

newListQueryExecutions :: ListQueryExecutions

Create a value of ListQueryExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListQueryExecutions', listQueryExecutions_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:maxResults:ListQueryExecutions', listQueryExecutions_maxResults - The maximum number of query executions to return in this request.

$sel:workGroup:ListQueryExecutions', listQueryExecutions_workGroup - The name of the workgroup from which queries are being returned. If a workgroup is not specified, a list of available query execution IDs for the queries in the primary workgroup is returned.

Request Lenses

listQueryExecutions_nextToken :: Lens' ListQueryExecutions (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listQueryExecutions_maxResults :: Lens' ListQueryExecutions (Maybe Natural)

The maximum number of query executions to return in this request.

listQueryExecutions_workGroup :: Lens' ListQueryExecutions (Maybe Text)

The name of the workgroup from which queries are being returned. If a workgroup is not specified, a list of available query execution IDs for the queries in the primary workgroup is returned.

Destructuring the Response

data ListQueryExecutionsResponse

See: newListQueryExecutionsResponse smart constructor.

Constructors

ListQueryExecutionsResponse'

Fields

nextToken :: Maybe Text

A token to be used by the next request if this request is truncated.

queryExecutionIds :: Maybe (NonEmpty Text)

The unique IDs of each query execution as an array of strings.

httpStatus :: Int

The response's http status code.

newListQueryExecutionsResponse

Arguments

:: Int
$sel:httpStatus:ListQueryExecutionsResponse'

-> ListQueryExecutionsResponse

Create a value of ListQueryExecutionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListQueryExecutions', listQueryExecutionsResponse_nextToken - A token to be used by the next request if this request is truncated.

$sel:queryExecutionIds:ListQueryExecutionsResponse', listQueryExecutionsResponse_queryExecutionIds - The unique IDs of each query execution as an array of strings.

$sel:httpStatus:ListQueryExecutionsResponse', listQueryExecutionsResponse_httpStatus - The response's http status code.

Response Lenses

listQueryExecutionsResponse_nextToken :: Lens' ListQueryExecutionsResponse (Maybe Text)

A token to be used by the next request if this request is truncated.

listQueryExecutionsResponse_queryExecutionIds :: Lens' ListQueryExecutionsResponse (Maybe (NonEmpty Text))

The unique IDs of each query execution as an array of strings.

listQueryExecutionsResponse_httpStatus :: Lens' ListQueryExecutionsResponse Int

The response's http status code.