Network.AWS.Athena.ListTableMetadata

Description

Lists the metadata for the tables in the specified data catalog database.

This operation returns paginated results.

Creating a Request

data ListTableMetadata

See: newListTableMetadata smart constructor.

Constructors

ListTableMetadata'

Fields

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

expression :: Maybe Text

A regex filter that pattern-matches table names. If no expression is supplied, metadata for all tables are listed.

maxResults :: Maybe Natural

Specifies the maximum number of results to return.

catalogName :: Text

The name of the data catalog for which table metadata should be returned.

databaseName :: Text

The name of the database for which table metadata should be returned.

newListTableMetadata

Arguments

:: Text
$sel:catalogName:ListTableMetadata'

-> Text
$sel:databaseName:ListTableMetadata'

-> ListTableMetadata

Create a value of ListTableMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTableMetadata', listTableMetadata_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:expression:ListTableMetadata', listTableMetadata_expression - A regex filter that pattern-matches table names. If no expression is supplied, metadata for all tables are listed.

$sel:maxResults:ListTableMetadata', listTableMetadata_maxResults - Specifies the maximum number of results to return.

$sel:catalogName:ListTableMetadata', listTableMetadata_catalogName - The name of the data catalog for which table metadata should be returned.

$sel:databaseName:ListTableMetadata', listTableMetadata_databaseName - The name of the database for which table metadata should be returned.

Request Lenses

listTableMetadata_nextToken :: Lens' ListTableMetadata (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listTableMetadata_expression :: Lens' ListTableMetadata (Maybe Text)

A regex filter that pattern-matches table names. If no expression is supplied, metadata for all tables are listed.

listTableMetadata_maxResults :: Lens' ListTableMetadata (Maybe Natural)

Specifies the maximum number of results to return.

listTableMetadata_catalogName :: Lens' ListTableMetadata Text

The name of the data catalog for which table metadata should be returned.

listTableMetadata_databaseName :: Lens' ListTableMetadata Text

The name of the database for which table metadata should be returned.

Destructuring the Response

data ListTableMetadataResponse

See: newListTableMetadataResponse smart constructor.

Constructors

ListTableMetadataResponse'

Fields

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

tableMetadataList :: Maybe [TableMetadata]

A list of table metadata.

httpStatus :: Int

The response's http status code.

newListTableMetadataResponse

Arguments

:: Int
$sel:httpStatus:ListTableMetadataResponse'

-> ListTableMetadataResponse

Create a value of ListTableMetadataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTableMetadata', listTableMetadataResponse_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:tableMetadataList:ListTableMetadataResponse', listTableMetadataResponse_tableMetadataList - A list of table metadata.

$sel:httpStatus:ListTableMetadataResponse', listTableMetadataResponse_httpStatus - The response's http status code.

Response Lenses

listTableMetadataResponse_nextToken :: Lens' ListTableMetadataResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listTableMetadataResponse_tableMetadataList :: Lens' ListTableMetadataResponse (Maybe [TableMetadata])

A list of table metadata.

listTableMetadataResponse_httpStatus :: Lens' ListTableMetadataResponse Int

The response's http status code.