Network.AWS.Athena.ListWorkGroups

Description

Lists available workgroups for the account.

Creating a Request

data ListWorkGroups

See: newListWorkGroups smart constructor.

Constructors

ListWorkGroups'

Fields

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

maxResults :: Maybe Natural

The maximum number of workgroups to return in this request.

newListWorkGroups :: ListWorkGroups

Create a value of ListWorkGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorkGroups', listWorkGroups_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:maxResults:ListWorkGroups', listWorkGroups_maxResults - The maximum number of workgroups to return in this request.

Request Lenses

listWorkGroups_nextToken :: Lens' ListWorkGroups (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listWorkGroups_maxResults :: Lens' ListWorkGroups (Maybe Natural)

The maximum number of workgroups to return in this request.

Destructuring the Response

data ListWorkGroupsResponse

See: newListWorkGroupsResponse smart constructor.

Constructors

ListWorkGroupsResponse'

Fields

nextToken :: Maybe Text

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

workGroups :: Maybe [WorkGroupSummary]

A list of WorkGroupSummary objects that include the names, descriptions, creation times, and states for each workgroup.

httpStatus :: Int

The response's http status code.

newListWorkGroupsResponse

Arguments

:: Int
$sel:httpStatus:ListWorkGroupsResponse'

-> ListWorkGroupsResponse

Create a value of ListWorkGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorkGroups', listWorkGroupsResponse_nextToken - A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

$sel:workGroups:ListWorkGroupsResponse', listWorkGroupsResponse_workGroups - A list of WorkGroupSummary objects that include the names, descriptions, creation times, and states for each workgroup.

$sel:httpStatus:ListWorkGroupsResponse', listWorkGroupsResponse_httpStatus - The response's http status code.

Response Lenses

listWorkGroupsResponse_nextToken :: Lens' ListWorkGroupsResponse (Maybe Text)

A token generated by the Athena service that specifies where to continue pagination if a previous request was truncated. To obtain the next set of pages, pass in the NextToken from the response object of the previous page call.

listWorkGroupsResponse_workGroups :: Lens' ListWorkGroupsResponse (Maybe [WorkGroupSummary])

A list of WorkGroupSummary objects that include the names, descriptions, creation times, and states for each workgroup.

listWorkGroupsResponse_httpStatus :: Lens' ListWorkGroupsResponse Int

The response's http status code.