Network.AWS.Athena.Types.DataCatalog

Description

Documentation

data DataCatalog

Contains information about a data catalog in an Amazon Web Services account.

See: newDataCatalog smart constructor.

Constructors

DataCatalog'

Fields

description :: Maybe Text

An optional description of the data catalog.

parameters :: Maybe (HashMap Text Text)

Specifies the Lambda function or functions to use for the data catalog. This is a mapping whose values depend on the catalog type.

For the HIVE data catalog type, use the following syntax. The metadata-function parameter is required. The sdk-version parameter is optional and defaults to the currently supported version.

metadata-function=lambda_arn, sdk-version=version_number

For the LAMBDA data catalog type, use one of the following sets of required parameters, but not both.

If you have one Lambda function that processes metadata and another for reading the actual data, use the following syntax. Both parameters are required.

metadata-function=lambda_arn, record-function=lambda_arn

If you have a composite Lambda function that processes both metadata and data, use the following syntax to specify your Lambda function.

function=lambda_arn

The GLUE type takes a catalog ID parameter and is required. The catalog_id is the account ID of the Amazon Web Services account to which the Glue catalog belongs.

catalog-id=catalog_id

The GLUE data catalog type also applies to the default AwsDataCatalog that already exists in your account, of which you can have only one and cannot modify.

Queries that specify a Glue Data Catalog other than the default AwsDataCatalog must be run on Athena engine version 2.

name :: Text

The name of the data catalog. The catalog name must be unique for the Amazon Web Services account and can use a maximum of 128 alphanumeric, underscore, at sign, or hyphen characters.

type' :: DataCatalogType

The type of data catalog to create: LAMBDA for a federated catalog, HIVE for an external hive metastore, or GLUE for an Glue Data Catalog.

newDataCatalog

Arguments

:: Text
$sel:name:DataCatalog'

-> DataCatalogType
$sel:type':DataCatalog'

-> DataCatalog

Create a value of DataCatalog with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:DataCatalog', dataCatalog_description - An optional description of the data catalog.

$sel:parameters:DataCatalog', dataCatalog_parameters - Specifies the Lambda function or functions to use for the data catalog. This is a mapping whose values depend on the catalog type.

For the HIVE data catalog type, use the following syntax. The metadata-function parameter is required. The sdk-version parameter is optional and defaults to the currently supported version.

metadata-function=lambda_arn, sdk-version=version_number

For the LAMBDA data catalog type, use one of the following sets of required parameters, but not both.

If you have one Lambda function that processes metadata and another for reading the actual data, use the following syntax. Both parameters are required.

metadata-function=lambda_arn, record-function=lambda_arn

If you have a composite Lambda function that processes both metadata and data, use the following syntax to specify your Lambda function.

function=lambda_arn

The GLUE type takes a catalog ID parameter and is required. The catalog_id is the account ID of the Amazon Web Services account to which the Glue catalog belongs.

catalog-id=catalog_id

The GLUE data catalog type also applies to the default AwsDataCatalog that already exists in your account, of which you can have only one and cannot modify.

Queries that specify a Glue Data Catalog other than the default AwsDataCatalog must be run on Athena engine version 2.

$sel:name:DataCatalog', dataCatalog_name - The name of the data catalog. The catalog name must be unique for the Amazon Web Services account and can use a maximum of 128 alphanumeric, underscore, at sign, or hyphen characters.

$sel:type':DataCatalog', dataCatalog_type - The type of data catalog to create: LAMBDA for a federated catalog, HIVE for an external hive metastore, or GLUE for an Glue Data Catalog.

dataCatalog_description :: Lens' DataCatalog (Maybe Text)

An optional description of the data catalog.

dataCatalog_parameters :: Lens' DataCatalog (Maybe (HashMap Text Text))

Specifies the Lambda function or functions to use for the data catalog. This is a mapping whose values depend on the catalog type.

For the HIVE data catalog type, use the following syntax. The metadata-function parameter is required. The sdk-version parameter is optional and defaults to the currently supported version.

metadata-function=lambda_arn, sdk-version=version_number

For the LAMBDA data catalog type, use one of the following sets of required parameters, but not both.

If you have one Lambda function that processes metadata and another for reading the actual data, use the following syntax. Both parameters are required.

metadata-function=lambda_arn, record-function=lambda_arn

If you have a composite Lambda function that processes both metadata and data, use the following syntax to specify your Lambda function.

function=lambda_arn

The GLUE type takes a catalog ID parameter and is required. The catalog_id is the account ID of the Amazon Web Services account to which the Glue catalog belongs.

catalog-id=catalog_id

The GLUE data catalog type also applies to the default AwsDataCatalog that already exists in your account, of which you can have only one and cannot modify.

Queries that specify a Glue Data Catalog other than the default AwsDataCatalog must be run on Athena engine version 2.

dataCatalog_name :: Lens' DataCatalog Text

The name of the data catalog. The catalog name must be unique for the Amazon Web Services account and can use a maximum of 128 alphanumeric, underscore, at sign, or hyphen characters.

dataCatalog_type :: Lens' DataCatalog DataCatalogType

The type of data catalog to create: LAMBDA for a federated catalog, HIVE for an external hive metastore, or GLUE for an Glue Data Catalog.