Network.AWS.Athena.Types.QueryExecution

Description

Documentation

data QueryExecution

Information about a single instance of a query execution.

See: newQueryExecution smart constructor.

Constructors

QueryExecution'

Fields

status :: Maybe QueryExecutionStatus

The completion date, current state, submission time, and state change reason (if applicable) for the query execution.

queryExecutionId :: Maybe Text

The unique identifier for each query execution.

statistics :: Maybe QueryExecutionStatistics

Query execution statistics, such as the amount of data scanned, the amount of time that the query took to process, and the type of statement that was run.

query :: Maybe Text

The SQL query statements which the query execution ran.

queryExecutionContext :: Maybe QueryExecutionContext

The database in which the query execution occurred.

engineVersion :: Maybe EngineVersion

The engine version that executed the query.

resultConfiguration :: Maybe ResultConfiguration

The location in Amazon S3 where query results were stored and the encryption option, if any, used for query results. These are known as "client-side settings". If workgroup settings override client-side settings, then the query uses the location for the query results and the encryption configuration that are specified for the workgroup.

workGroup :: Maybe Text

The name of the workgroup in which the query ran.

statementType :: Maybe StatementType

The type of query statement that was run. DDL indicates DDL query statements. DML indicates DML (Data Manipulation Language) query statements, such as CREATE TABLE AS SELECT. UTILITY indicates query statements other than DDL and DML, such as SHOW CREATE TABLE, or DESCRIBE TABLE.

newQueryExecution :: QueryExecution

Create a value of QueryExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:QueryExecution', queryExecution_status - The completion date, current state, submission time, and state change reason (if applicable) for the query execution.

$sel:queryExecutionId:QueryExecution', queryExecution_queryExecutionId - The unique identifier for each query execution.

$sel:statistics:QueryExecution', queryExecution_statistics - Query execution statistics, such as the amount of data scanned, the amount of time that the query took to process, and the type of statement that was run.

$sel:query:QueryExecution', queryExecution_query - The SQL query statements which the query execution ran.

$sel:queryExecutionContext:QueryExecution', queryExecution_queryExecutionContext - The database in which the query execution occurred.

$sel:engineVersion:QueryExecution', queryExecution_engineVersion - The engine version that executed the query.

$sel:resultConfiguration:QueryExecution', queryExecution_resultConfiguration - The location in Amazon S3 where query results were stored and the encryption option, if any, used for query results. These are known as "client-side settings". If workgroup settings override client-side settings, then the query uses the location for the query results and the encryption configuration that are specified for the workgroup.

$sel:workGroup:QueryExecution', queryExecution_workGroup - The name of the workgroup in which the query ran.

$sel:statementType:QueryExecution', queryExecution_statementType - The type of query statement that was run. DDL indicates DDL query statements. DML indicates DML (Data Manipulation Language) query statements, such as CREATE TABLE AS SELECT. UTILITY indicates query statements other than DDL and DML, such as SHOW CREATE TABLE, or DESCRIBE TABLE.

queryExecution_status :: Lens' QueryExecution (Maybe QueryExecutionStatus)

The completion date, current state, submission time, and state change reason (if applicable) for the query execution.

queryExecution_queryExecutionId :: Lens' QueryExecution (Maybe Text)

The unique identifier for each query execution.

queryExecution_statistics :: Lens' QueryExecution (Maybe QueryExecutionStatistics)

Query execution statistics, such as the amount of data scanned, the amount of time that the query took to process, and the type of statement that was run.

queryExecution_query :: Lens' QueryExecution (Maybe Text)

The SQL query statements which the query execution ran.

queryExecution_queryExecutionContext :: Lens' QueryExecution (Maybe QueryExecutionContext)

The database in which the query execution occurred.

queryExecution_engineVersion :: Lens' QueryExecution (Maybe EngineVersion)

The engine version that executed the query.

queryExecution_resultConfiguration :: Lens' QueryExecution (Maybe ResultConfiguration)

The location in Amazon S3 where query results were stored and the encryption option, if any, used for query results. These are known as "client-side settings". If workgroup settings override client-side settings, then the query uses the location for the query results and the encryption configuration that are specified for the workgroup.

queryExecution_workGroup :: Lens' QueryExecution (Maybe Text)

The name of the workgroup in which the query ran.

queryExecution_statementType :: Lens' QueryExecution (Maybe StatementType)

The type of query statement that was run. DDL indicates DDL query statements. DML indicates DML (Data Manipulation Language) query statements, such as CREATE TABLE AS SELECT. UTILITY indicates query statements other than DDL and DML, such as SHOW CREATE TABLE, or DESCRIBE TABLE.