Network.AWS.Athena.Types.ResultSet

Description

Documentation

data ResultSet

The metadata and rows that comprise a query result set. The metadata describes the column structure and data types. To return a ResultSet object, use GetQueryResults.

See: newResultSet smart constructor.

Constructors

ResultSet'

Fields

rows :: Maybe [Row]

The rows in the table.

resultSetMetadata :: Maybe ResultSetMetadata

The metadata that describes the column structure and data types of a table of query results.

newResultSet :: ResultSet

Create a value of ResultSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rows:ResultSet', resultSet_rows - The rows in the table.

$sel:resultSetMetadata:ResultSet', resultSet_resultSetMetadata - The metadata that describes the column structure and data types of a table of query results.

resultSet_rows :: Lens' ResultSet (Maybe [Row])

The rows in the table.

resultSet_resultSetMetadata :: Lens' ResultSet (Maybe ResultSetMetadata)

The metadata that describes the column structure and data types of a table of query results.