Network.AWS.Athena.Types.WorkGroup

Description

Documentation

data WorkGroup

A workgroup, which contains a name, description, creation time, state, and other configuration, listed under WorkGroup$Configuration. Each workgroup enables you to isolate queries for you or your group of users from other queries in the same account, to configure the query results location and the encryption configuration (known as workgroup settings), to enable sending query metrics to Amazon CloudWatch, and to establish per-query data usage control limits for all queries in a workgroup. The workgroup settings override is specified in EnforceWorkGroupConfiguration (true/false) in the WorkGroupConfiguration. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

See: newWorkGroup smart constructor.

Constructors

WorkGroup'

Fields

creationTime :: Maybe POSIX

The date and time the workgroup was created.

configuration :: Maybe WorkGroupConfiguration

The configuration of the workgroup, which includes the location in Amazon S3 where query results are stored, the encryption configuration, if any, used for query results; whether the Amazon CloudWatch Metrics are enabled for the workgroup; whether workgroup settings override client-side settings; and the data usage limits for the amount of data scanned per query or per workgroup. The workgroup settings override is specified in EnforceWorkGroupConfiguration (true/false) in the WorkGroupConfiguration. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

state :: Maybe WorkGroupState

The state of the workgroup: ENABLED or DISABLED.

description :: Maybe Text

The workgroup description.

name :: Text

The workgroup name.

newWorkGroup

Arguments

:: Text
$sel:name:WorkGroup'

-> WorkGroup

Create a value of WorkGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:WorkGroup', workGroup_creationTime - The date and time the workgroup was created.

$sel:configuration:WorkGroup', workGroup_configuration - The configuration of the workgroup, which includes the location in Amazon S3 where query results are stored, the encryption configuration, if any, used for query results; whether the Amazon CloudWatch Metrics are enabled for the workgroup; whether workgroup settings override client-side settings; and the data usage limits for the amount of data scanned per query or per workgroup. The workgroup settings override is specified in EnforceWorkGroupConfiguration (true/false) in the WorkGroupConfiguration. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

$sel:state:WorkGroup', workGroup_state - The state of the workgroup: ENABLED or DISABLED.

$sel:description:WorkGroup', workGroup_description - The workgroup description.

$sel:name:WorkGroup', workGroup_name - The workgroup name.

workGroup_creationTime :: Lens' WorkGroup (Maybe UTCTime)

The date and time the workgroup was created.

workGroup_configuration :: Lens' WorkGroup (Maybe WorkGroupConfiguration)

The configuration of the workgroup, which includes the location in Amazon S3 where query results are stored, the encryption configuration, if any, used for query results; whether the Amazon CloudWatch Metrics are enabled for the workgroup; whether workgroup settings override client-side settings; and the data usage limits for the amount of data scanned per query or per workgroup. The workgroup settings override is specified in EnforceWorkGroupConfiguration (true/false) in the WorkGroupConfiguration. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

workGroup_state :: Lens' WorkGroup (Maybe WorkGroupState)

The state of the workgroup: ENABLED or DISABLED.

workGroup_description :: Lens' WorkGroup (Maybe Text)

The workgroup description.

workGroup_name :: Lens' WorkGroup Text

The workgroup name.