Network.AWS.Athena.Types.WorkGroupConfiguration

Description

Documentation

data WorkGroupConfiguration

The configuration of the workgroup, which includes the location in Amazon S3 where query results are stored, the encryption option, if any, used for query results, whether the Amazon CloudWatch Metrics are enabled for the workgroup and whether workgroup settings override query settings, and the data usage limits for the amount of data scanned per query or per workgroup. The workgroup settings override is specified in EnforceWorkGroupConfiguration (true/false) in the WorkGroupConfiguration. See WorkGroupConfiguration$EnforceWorkGroupConfiguration.

See: newWorkGroupConfiguration smart constructor.

Constructors

WorkGroupConfiguration'

Fields

bytesScannedCutoffPerQuery :: Maybe Natural

The upper data usage limit (cutoff) for the amount of bytes a single query in a workgroup is allowed to scan.

publishCloudWatchMetricsEnabled :: Maybe Bool

Indicates that the Amazon CloudWatch metrics are enabled for the workgroup.

requesterPaysEnabled :: Maybe Bool

If set to true, allows members assigned to a workgroup to reference Amazon S3 Requester Pays buckets in queries. If set to false, workgroup members cannot query data from Requester Pays buckets, and queries that retrieve data from Requester Pays buckets cause an error. The default is false. For more information about Requester Pays buckets, see Requester Pays Buckets in the Amazon Simple Storage Service Developer Guide.

enforceWorkGroupConfiguration :: Maybe Bool

If set to "true", the settings for the workgroup override client-side settings. If set to "false", client-side settings are used. For more information, see Workgroup Settings Override Client-Side Settings.

engineVersion :: Maybe EngineVersion

The engine version that all queries running on the workgroup use. Queries on the AmazonAthenaPreviewFunctionality workgroup run on the preview engine regardless of this setting.

resultConfiguration :: Maybe ResultConfiguration

The configuration for the workgroup, which includes the location in Amazon S3 where query results are stored and the encryption option, if any, used for query results. To run the query, you must specify the query results location using one of the ways: either in the workgroup using this setting, or for individual queries (client-side), using ResultConfiguration$OutputLocation. If none of them is set, Athena issues an error that no output location is provided. For more information, see Query Results.

newWorkGroupConfiguration :: WorkGroupConfiguration

Create a value of WorkGroupConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bytesScannedCutoffPerQuery:WorkGroupConfiguration', workGroupConfiguration_bytesScannedCutoffPerQuery - The upper data usage limit (cutoff) for the amount of bytes a single query in a workgroup is allowed to scan.

$sel:publishCloudWatchMetricsEnabled:WorkGroupConfiguration', workGroupConfiguration_publishCloudWatchMetricsEnabled - Indicates that the Amazon CloudWatch metrics are enabled for the workgroup.

$sel:requesterPaysEnabled:WorkGroupConfiguration', workGroupConfiguration_requesterPaysEnabled - If set to true, allows members assigned to a workgroup to reference Amazon S3 Requester Pays buckets in queries. If set to false, workgroup members cannot query data from Requester Pays buckets, and queries that retrieve data from Requester Pays buckets cause an error. The default is false. For more information about Requester Pays buckets, see Requester Pays Buckets in the Amazon Simple Storage Service Developer Guide.

$sel:enforceWorkGroupConfiguration:WorkGroupConfiguration', workGroupConfiguration_enforceWorkGroupConfiguration - If set to "true", the settings for the workgroup override client-side settings. If set to "false", client-side settings are used. For more information, see Workgroup Settings Override Client-Side Settings.

$sel:engineVersion:WorkGroupConfiguration', workGroupConfiguration_engineVersion - The engine version that all queries running on the workgroup use. Queries on the AmazonAthenaPreviewFunctionality workgroup run on the preview engine regardless of this setting.

$sel:resultConfiguration:WorkGroupConfiguration', workGroupConfiguration_resultConfiguration - The configuration for the workgroup, which includes the location in Amazon S3 where query results are stored and the encryption option, if any, used for query results. To run the query, you must specify the query results location using one of the ways: either in the workgroup using this setting, or for individual queries (client-side), using ResultConfiguration$OutputLocation. If none of them is set, Athena issues an error that no output location is provided. For more information, see Query Results.

workGroupConfiguration_bytesScannedCutoffPerQuery :: Lens' WorkGroupConfiguration (Maybe Natural)

The upper data usage limit (cutoff) for the amount of bytes a single query in a workgroup is allowed to scan.

workGroupConfiguration_publishCloudWatchMetricsEnabled :: Lens' WorkGroupConfiguration (Maybe Bool)

Indicates that the Amazon CloudWatch metrics are enabled for the workgroup.

workGroupConfiguration_requesterPaysEnabled :: Lens' WorkGroupConfiguration (Maybe Bool)

If set to true, allows members assigned to a workgroup to reference Amazon S3 Requester Pays buckets in queries. If set to false, workgroup members cannot query data from Requester Pays buckets, and queries that retrieve data from Requester Pays buckets cause an error. The default is false. For more information about Requester Pays buckets, see Requester Pays Buckets in the Amazon Simple Storage Service Developer Guide.

workGroupConfiguration_enforceWorkGroupConfiguration :: Lens' WorkGroupConfiguration (Maybe Bool)

If set to "true", the settings for the workgroup override client-side settings. If set to "false", client-side settings are used. For more information, see Workgroup Settings Override Client-Side Settings.

workGroupConfiguration_engineVersion :: Lens' WorkGroupConfiguration (Maybe EngineVersion)

The engine version that all queries running on the workgroup use. Queries on the AmazonAthenaPreviewFunctionality workgroup run on the preview engine regardless of this setting.

workGroupConfiguration_resultConfiguration :: Lens' WorkGroupConfiguration (Maybe ResultConfiguration)

The configuration for the workgroup, which includes the location in Amazon S3 where query results are stored and the encryption option, if any, used for query results. To run the query, you must specify the query results location using one of the ways: either in the workgroup using this setting, or for individual queries (client-side), using ResultConfiguration$OutputLocation. If none of them is set, Athena issues an error that no output location is provided. For more information, see Query Results.