Network.AWS.Athena.Types.WorkGroupConfigurationUpdates

Description

Documentation

data WorkGroupConfigurationUpdates

The configuration information that will be updated for this workgroup, which includes the location in Amazon S3 where query results are stored, the encryption option, if any, used for query results, whether the Amazon CloudWatch Metrics are enabled for the workgroup, whether the workgroup settings override the client-side settings, and the data usage limit for the amount of bytes scanned per query, if it is specified.

See: newWorkGroupConfigurationUpdates smart constructor.

Constructors

WorkGroupConfigurationUpdates'

Fields

bytesScannedCutoffPerQuery :: Maybe Natural

The upper limit (cutoff) for the amount of bytes a single query in a workgroup is allowed to scan.

publishCloudWatchMetricsEnabled :: Maybe Bool

Indicates whether this workgroup enables publishing metrics to Amazon CloudWatch.

resultConfigurationUpdates :: Maybe ResultConfigurationUpdates

The result configuration information about the queries in this workgroup that will be updated. Includes the updated results location and an updated option for encrypting query results.

requesterPaysEnabled :: Maybe Bool

If set to true, allows members assigned to a workgroup to specify Amazon S3 Requester Pays buckets in queries. If set to false, workgroup members cannot query data from Requester Pays buckets, and queries that retrieve data from Requester Pays buckets cause an error. The default is false. For more information about Requester Pays buckets, see Requester Pays Buckets in the Amazon Simple Storage Service Developer Guide.

enforceWorkGroupConfiguration :: Maybe Bool

If set to "true", the settings for the workgroup override client-side settings. If set to "false" client-side settings are used. For more information, see Workgroup Settings Override Client-Side Settings.

removeBytesScannedCutoffPerQuery :: Maybe Bool

Indicates that the data usage control limit per query is removed. WorkGroupConfiguration$BytesScannedCutoffPerQuery

engineVersion :: Maybe EngineVersion

The engine version requested when a workgroup is updated. After the update, all queries on the workgroup run on the requested engine version. If no value was previously set, the default is Auto. Queries on the AmazonAthenaPreviewFunctionality workgroup run on the preview engine regardless of this setting.

newWorkGroupConfigurationUpdates :: WorkGroupConfigurationUpdates

Create a value of WorkGroupConfigurationUpdates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bytesScannedCutoffPerQuery:WorkGroupConfigurationUpdates', workGroupConfigurationUpdates_bytesScannedCutoffPerQuery - The upper limit (cutoff) for the amount of bytes a single query in a workgroup is allowed to scan.

$sel:publishCloudWatchMetricsEnabled:WorkGroupConfigurationUpdates', workGroupConfigurationUpdates_publishCloudWatchMetricsEnabled - Indicates whether this workgroup enables publishing metrics to Amazon CloudWatch.

$sel:resultConfigurationUpdates:WorkGroupConfigurationUpdates', workGroupConfigurationUpdates_resultConfigurationUpdates - The result configuration information about the queries in this workgroup that will be updated. Includes the updated results location and an updated option for encrypting query results.

$sel:requesterPaysEnabled:WorkGroupConfigurationUpdates', workGroupConfigurationUpdates_requesterPaysEnabled - If set to true, allows members assigned to a workgroup to specify Amazon S3 Requester Pays buckets in queries. If set to false, workgroup members cannot query data from Requester Pays buckets, and queries that retrieve data from Requester Pays buckets cause an error. The default is false. For more information about Requester Pays buckets, see Requester Pays Buckets in the Amazon Simple Storage Service Developer Guide.

$sel:enforceWorkGroupConfiguration:WorkGroupConfigurationUpdates', workGroupConfigurationUpdates_enforceWorkGroupConfiguration - If set to "true", the settings for the workgroup override client-side settings. If set to "false" client-side settings are used. For more information, see Workgroup Settings Override Client-Side Settings.

$sel:removeBytesScannedCutoffPerQuery:WorkGroupConfigurationUpdates', workGroupConfigurationUpdates_removeBytesScannedCutoffPerQuery - Indicates that the data usage control limit per query is removed. WorkGroupConfiguration$BytesScannedCutoffPerQuery

$sel:engineVersion:WorkGroupConfigurationUpdates', workGroupConfigurationUpdates_engineVersion - The engine version requested when a workgroup is updated. After the update, all queries on the workgroup run on the requested engine version. If no value was previously set, the default is Auto. Queries on the AmazonAthenaPreviewFunctionality workgroup run on the preview engine regardless of this setting.

workGroupConfigurationUpdates_bytesScannedCutoffPerQuery :: Lens' WorkGroupConfigurationUpdates (Maybe Natural)

The upper limit (cutoff) for the amount of bytes a single query in a workgroup is allowed to scan.

workGroupConfigurationUpdates_publishCloudWatchMetricsEnabled :: Lens' WorkGroupConfigurationUpdates (Maybe Bool)

Indicates whether this workgroup enables publishing metrics to Amazon CloudWatch.

workGroupConfigurationUpdates_resultConfigurationUpdates :: Lens' WorkGroupConfigurationUpdates (Maybe ResultConfigurationUpdates)

The result configuration information about the queries in this workgroup that will be updated. Includes the updated results location and an updated option for encrypting query results.

workGroupConfigurationUpdates_requesterPaysEnabled :: Lens' WorkGroupConfigurationUpdates (Maybe Bool)

If set to true, allows members assigned to a workgroup to specify Amazon S3 Requester Pays buckets in queries. If set to false, workgroup members cannot query data from Requester Pays buckets, and queries that retrieve data from Requester Pays buckets cause an error. The default is false. For more information about Requester Pays buckets, see Requester Pays Buckets in the Amazon Simple Storage Service Developer Guide.

workGroupConfigurationUpdates_enforceWorkGroupConfiguration :: Lens' WorkGroupConfigurationUpdates (Maybe Bool)

If set to "true", the settings for the workgroup override client-side settings. If set to "false" client-side settings are used. For more information, see Workgroup Settings Override Client-Side Settings.

workGroupConfigurationUpdates_removeBytesScannedCutoffPerQuery :: Lens' WorkGroupConfigurationUpdates (Maybe Bool)

Indicates that the data usage control limit per query is removed. WorkGroupConfiguration$BytesScannedCutoffPerQuery

workGroupConfigurationUpdates_engineVersion :: Lens' WorkGroupConfigurationUpdates (Maybe EngineVersion)

The engine version requested when a workgroup is updated. After the update, all queries on the workgroup run on the requested engine version. If no value was previously set, the default is Auto. Queries on the AmazonAthenaPreviewFunctionality workgroup run on the preview engine regardless of this setting.