Network.AWS.AuditManager.DeleteAssessmentFramework

Description

Deletes a custom framework in Audit Manager.

Creating a Request

data DeleteAssessmentFramework

See: newDeleteAssessmentFramework smart constructor.

Constructors

DeleteAssessmentFramework'

Fields

frameworkId :: Text

The identifier for the specified framework.

newDeleteAssessmentFramework

Arguments

:: Text
$sel:frameworkId:DeleteAssessmentFramework'

-> DeleteAssessmentFramework

Create a value of DeleteAssessmentFramework with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:frameworkId:DeleteAssessmentFramework', deleteAssessmentFramework_frameworkId - The identifier for the specified framework.

Request Lenses

deleteAssessmentFramework_frameworkId :: Lens' DeleteAssessmentFramework Text

The identifier for the specified framework.

Destructuring the Response

data DeleteAssessmentFrameworkResponse

See: newDeleteAssessmentFrameworkResponse smart constructor.

Constructors

DeleteAssessmentFrameworkResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteAssessmentFrameworkResponse

Arguments

:: Int
$sel:httpStatus:DeleteAssessmentFrameworkResponse'

-> DeleteAssessmentFrameworkResponse

Create a value of DeleteAssessmentFrameworkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteAssessmentFrameworkResponse', deleteAssessmentFrameworkResponse_httpStatus - The response's http status code.

Response Lenses

deleteAssessmentFrameworkResponse_httpStatus :: Lens' DeleteAssessmentFrameworkResponse Int

The response's http status code.