Network.AWS.AuditManager.GetServicesInScope

Description

Returns a list of the in-scope Amazon Web Services services for the specified assessment.

Creating a Request

data GetServicesInScope

See: newGetServicesInScope smart constructor.

Constructors

GetServicesInScope'

newGetServicesInScope :: GetServicesInScope

Create a value of GetServicesInScope with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetServicesInScopeResponse

See: newGetServicesInScopeResponse smart constructor.

Constructors

GetServicesInScopeResponse'

Fields

serviceMetadata :: Maybe [ServiceMetadata]

The metadata associated with the Amazon Web Service.

httpStatus :: Int

The response's http status code.

newGetServicesInScopeResponse

Arguments

:: Int
$sel:httpStatus:GetServicesInScopeResponse'

-> GetServicesInScopeResponse

Create a value of GetServicesInScopeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceMetadata:GetServicesInScopeResponse', getServicesInScopeResponse_serviceMetadata - The metadata associated with the Amazon Web Service.

$sel:httpStatus:GetServicesInScopeResponse', getServicesInScopeResponse_httpStatus - The response's http status code.

Response Lenses

getServicesInScopeResponse_serviceMetadata :: Lens' GetServicesInScopeResponse (Maybe [ServiceMetadata])

The metadata associated with the Amazon Web Service.

getServicesInScopeResponse_httpStatus :: Lens' GetServicesInScopeResponse Int

The response's http status code.