Network.AWS.AuditManager.ListControls

Description

Returns a list of controls from Audit Manager.

Creating a Request

data ListControls

See: newListControls smart constructor.

Constructors

ListControls'

Fields

nextToken :: Maybe Text

The pagination token used to fetch the next set of results.

maxResults :: Maybe Natural

Represents the maximum number of results per page, or per API request call.

controlType :: ControlType

The type of control, such as standard or custom.

newListControls

Arguments

:: ControlType
$sel:controlType:ListControls'

-> ListControls

Create a value of ListControls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListControls', listControls_nextToken - The pagination token used to fetch the next set of results.

$sel:maxResults:ListControls', listControls_maxResults - Represents the maximum number of results per page, or per API request call.

$sel:controlType:ListControls', listControls_controlType - The type of control, such as standard or custom.

Request Lenses

listControls_nextToken :: Lens' ListControls (Maybe Text)

The pagination token used to fetch the next set of results.

listControls_maxResults :: Lens' ListControls (Maybe Natural)

Represents the maximum number of results per page, or per API request call.

listControls_controlType :: Lens' ListControls ControlType

The type of control, such as standard or custom.

Destructuring the Response

data ListControlsResponse

See: newListControlsResponse smart constructor.

Constructors

ListControlsResponse'

Fields

nextToken :: Maybe Text

The pagination token used to fetch the next set of results.

controlMetadataList :: Maybe [ControlMetadata]

The list of control metadata objects returned by the ListControls API.

httpStatus :: Int

The response's http status code.

newListControlsResponse

Arguments

:: Int
$sel:httpStatus:ListControlsResponse'

-> ListControlsResponse

Create a value of ListControlsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListControls', listControlsResponse_nextToken - The pagination token used to fetch the next set of results.

$sel:controlMetadataList:ListControlsResponse', listControlsResponse_controlMetadataList - The list of control metadata objects returned by the ListControls API.

$sel:httpStatus:ListControlsResponse', listControlsResponse_httpStatus - The response's http status code.

Response Lenses

listControlsResponse_nextToken :: Lens' ListControlsResponse (Maybe Text)

The pagination token used to fetch the next set of results.

listControlsResponse_controlMetadataList :: Lens' ListControlsResponse (Maybe [ControlMetadata])

The list of control metadata objects returned by the ListControls API.

listControlsResponse_httpStatus :: Lens' ListControlsResponse Int

The response's http status code.