Network.AWS.AuditManager.RegisterAccount

Description

Enables Audit Manager for the specified Amazon Web Services account.

Creating a Request

data RegisterAccount

See: newRegisterAccount smart constructor.

Constructors

RegisterAccount'

Fields

kmsKey :: Maybe Text

The KMS key details.

delegatedAdminAccount :: Maybe Text

The delegated administrator account for Audit Manager.

newRegisterAccount :: RegisterAccount

Create a value of RegisterAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKey:RegisterAccount', registerAccount_kmsKey - The KMS key details.

$sel:delegatedAdminAccount:RegisterAccount', registerAccount_delegatedAdminAccount - The delegated administrator account for Audit Manager.

Request Lenses

registerAccount_kmsKey :: Lens' RegisterAccount (Maybe Text)

The KMS key details.

registerAccount_delegatedAdminAccount :: Lens' RegisterAccount (Maybe Text)

The delegated administrator account for Audit Manager.

Destructuring the Response

data RegisterAccountResponse

See: newRegisterAccountResponse smart constructor.

Constructors

RegisterAccountResponse'

Fields

status :: Maybe AccountStatus

The status of the account registration request.

httpStatus :: Int

The response's http status code.

newRegisterAccountResponse

Arguments

:: Int
$sel:httpStatus:RegisterAccountResponse'

-> RegisterAccountResponse

Create a value of RegisterAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:RegisterAccountResponse', registerAccountResponse_status - The status of the account registration request.

$sel:httpStatus:RegisterAccountResponse', registerAccountResponse_httpStatus - The response's http status code.

Response Lenses

registerAccountResponse_status :: Lens' RegisterAccountResponse (Maybe AccountStatus)

The status of the account registration request.

registerAccountResponse_httpStatus :: Lens' RegisterAccountResponse Int

The response's http status code.