Network.AWS.AuditManager.RegisterOrganizationAdminAccount

Description

Enables an Amazon Web Services account within the organization as the delegated administrator for Audit Manager.

Creating a Request

data RegisterOrganizationAdminAccount

See: newRegisterOrganizationAdminAccount smart constructor.

Constructors

RegisterOrganizationAdminAccount'

Fields

adminAccountId :: Text

The identifier for the specified delegated administrator account.

newRegisterOrganizationAdminAccount

Arguments

:: Text
$sel:adminAccountId:RegisterOrganizationAdminAccount'

-> RegisterOrganizationAdminAccount

Create a value of RegisterOrganizationAdminAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:adminAccountId:RegisterOrganizationAdminAccount', registerOrganizationAdminAccount_adminAccountId - The identifier for the specified delegated administrator account.

Request Lenses

registerOrganizationAdminAccount_adminAccountId :: Lens' RegisterOrganizationAdminAccount Text

The identifier for the specified delegated administrator account.

Destructuring the Response

data RegisterOrganizationAdminAccountResponse

See: newRegisterOrganizationAdminAccountResponse smart constructor.

Constructors

RegisterOrganizationAdminAccountResponse'

Fields

adminAccountId :: Maybe Text

The identifier for the specified delegated administrator account.

organizationId :: Maybe Text

The identifier for the specified organization.

httpStatus :: Int

The response's http status code.

newRegisterOrganizationAdminAccountResponse

Arguments

:: Int
$sel:httpStatus:RegisterOrganizationAdminAccountResponse'

-> RegisterOrganizationAdminAccountResponse

Create a value of RegisterOrganizationAdminAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:adminAccountId:RegisterOrganizationAdminAccount', registerOrganizationAdminAccountResponse_adminAccountId - The identifier for the specified delegated administrator account.

$sel:organizationId:RegisterOrganizationAdminAccountResponse', registerOrganizationAdminAccountResponse_organizationId - The identifier for the specified organization.

$sel:httpStatus:RegisterOrganizationAdminAccountResponse', registerOrganizationAdminAccountResponse_httpStatus - The response's http status code.

Response Lenses

registerOrganizationAdminAccountResponse_adminAccountId :: Lens' RegisterOrganizationAdminAccountResponse (Maybe Text)

The identifier for the specified delegated administrator account.

registerOrganizationAdminAccountResponse_organizationId :: Lens' RegisterOrganizationAdminAccountResponse (Maybe Text)

The identifier for the specified organization.

registerOrganizationAdminAccountResponse_httpStatus :: Lens' RegisterOrganizationAdminAccountResponse Int

The response's http status code.