Network.AWS.AuditManager.Types.Assessment

Description

Documentation

data Assessment

An entity that defines the scope of audit evidence collected by Audit Manager. An Audit Manager assessment is an implementation of an Audit Manager framework.

See: newAssessment smart constructor.

Constructors

Assessment'

Fields

framework :: Maybe AssessmentFramework

The framework from which the assessment was created.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the assessment.

awsAccount :: Maybe AWSAccount

The Amazon Web Services account associated with the assessment.

metadata :: Maybe AssessmentMetadata

The metadata for the specified assessment.

tags :: Maybe (HashMap Text Text)

The tags associated with the assessment.

newAssessment :: Assessment

Create a value of Assessment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:framework:Assessment', assessment_framework - The framework from which the assessment was created.

$sel:arn:Assessment', assessment_arn - The Amazon Resource Name (ARN) of the assessment.

$sel:awsAccount:Assessment', assessment_awsAccount - The Amazon Web Services account associated with the assessment.

$sel:metadata:Assessment', assessment_metadata - The metadata for the specified assessment.

$sel:tags:Assessment', assessment_tags - The tags associated with the assessment.

assessment_framework :: Lens' Assessment (Maybe AssessmentFramework)

The framework from which the assessment was created.

assessment_arn :: Lens' Assessment (Maybe Text)

The Amazon Resource Name (ARN) of the assessment.

assessment_awsAccount :: Lens' Assessment (Maybe AWSAccount)

The Amazon Web Services account associated with the assessment.

assessment_metadata :: Lens' Assessment (Maybe AssessmentMetadata)

The metadata for the specified assessment.

assessment_tags :: Lens' Assessment (Maybe (HashMap Text Text))

The tags associated with the assessment.