Network.AWS.AuditManager.Types.AssessmentEvidenceFolder

Description

Documentation

data AssessmentEvidenceFolder

The folder in which Audit Manager stores evidence for an assessment.

See: newAssessmentEvidenceFolder smart constructor.

Constructors

AssessmentEvidenceFolder'

Fields

evidenceByTypeComplianceCheckIssuesCount :: Maybe Int

The total number of issues that were reported directly from Security Hub, Config, or both.

controlSetId :: Maybe Text

The identifier for the control set.

assessmentReportSelectionCount :: Maybe Int

The total count of evidence included in the assessment report.

totalEvidence :: Maybe Int

The total amount of evidence in the evidence folder.

evidenceByTypeManualCount :: Maybe Int

The number of evidence that falls under the manual category. This evidence is imported manually.

date :: Maybe POSIX

The date when the first evidence was added to the evidence folder.

name :: Maybe Text

The name of the specified evidence folder.

evidenceByTypeUserActivityCount :: Maybe Int

The number of evidence that falls under the user activity category. This evidence is collected from CloudTrail logs.

controlId :: Maybe Text

The unique identifier for the specified control.

evidenceAwsServiceSourceCount :: Maybe Int

The total number of Amazon Web Services resources assessed to generate the evidence.

author :: Maybe Text

The name of the user who created the evidence folder.

id :: Maybe Text

The identifier for the folder in which evidence is stored.

dataSource :: Maybe Text

The Amazon Web Service from which the evidence was collected.

controlName :: Maybe Text

The name of the control.

evidenceByTypeComplianceCheckCount :: Maybe Int

The number of evidence that falls under the compliance check category. This evidence is collected from Config or Security Hub.

assessmentId :: Maybe Text

The identifier for the specified assessment.

evidenceByTypeConfigurationDataCount :: Maybe Int

The number of evidence that falls under the configuration data category. This evidence is collected from configuration snapshots of other Amazon Web Services services such as Amazon EC2, Amazon S3, or IAM.

evidenceResourcesIncludedCount :: Maybe Int

The amount of evidence included in the evidence folder.

newAssessmentEvidenceFolder :: AssessmentEvidenceFolder

Create a value of AssessmentEvidenceFolder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:evidenceByTypeComplianceCheckIssuesCount:AssessmentEvidenceFolder', assessmentEvidenceFolder_evidenceByTypeComplianceCheckIssuesCount - The total number of issues that were reported directly from Security Hub, Config, or both.

$sel:controlSetId:AssessmentEvidenceFolder', assessmentEvidenceFolder_controlSetId - The identifier for the control set.

$sel:assessmentReportSelectionCount:AssessmentEvidenceFolder', assessmentEvidenceFolder_assessmentReportSelectionCount - The total count of evidence included in the assessment report.

$sel:totalEvidence:AssessmentEvidenceFolder', assessmentEvidenceFolder_totalEvidence - The total amount of evidence in the evidence folder.

$sel:evidenceByTypeManualCount:AssessmentEvidenceFolder', assessmentEvidenceFolder_evidenceByTypeManualCount - The number of evidence that falls under the manual category. This evidence is imported manually.

$sel:date:AssessmentEvidenceFolder', assessmentEvidenceFolder_date - The date when the first evidence was added to the evidence folder.

$sel:name:AssessmentEvidenceFolder', assessmentEvidenceFolder_name - The name of the specified evidence folder.

$sel:evidenceByTypeUserActivityCount:AssessmentEvidenceFolder', assessmentEvidenceFolder_evidenceByTypeUserActivityCount - The number of evidence that falls under the user activity category. This evidence is collected from CloudTrail logs.

$sel:controlId:AssessmentEvidenceFolder', assessmentEvidenceFolder_controlId - The unique identifier for the specified control.

$sel:evidenceAwsServiceSourceCount:AssessmentEvidenceFolder', assessmentEvidenceFolder_evidenceAwsServiceSourceCount - The total number of Amazon Web Services resources assessed to generate the evidence.

$sel:author:AssessmentEvidenceFolder', assessmentEvidenceFolder_author - The name of the user who created the evidence folder.

$sel:id:AssessmentEvidenceFolder', assessmentEvidenceFolder_id - The identifier for the folder in which evidence is stored.

$sel:dataSource:AssessmentEvidenceFolder', assessmentEvidenceFolder_dataSource - The Amazon Web Service from which the evidence was collected.

$sel:controlName:AssessmentEvidenceFolder', assessmentEvidenceFolder_controlName - The name of the control.

$sel:evidenceByTypeComplianceCheckCount:AssessmentEvidenceFolder', assessmentEvidenceFolder_evidenceByTypeComplianceCheckCount - The number of evidence that falls under the compliance check category. This evidence is collected from Config or Security Hub.

$sel:assessmentId:AssessmentEvidenceFolder', assessmentEvidenceFolder_assessmentId - The identifier for the specified assessment.

$sel:evidenceByTypeConfigurationDataCount:AssessmentEvidenceFolder', assessmentEvidenceFolder_evidenceByTypeConfigurationDataCount - The number of evidence that falls under the configuration data category. This evidence is collected from configuration snapshots of other Amazon Web Services services such as Amazon EC2, Amazon S3, or IAM.

$sel:evidenceResourcesIncludedCount:AssessmentEvidenceFolder', assessmentEvidenceFolder_evidenceResourcesIncludedCount - The amount of evidence included in the evidence folder.

assessmentEvidenceFolder_evidenceByTypeComplianceCheckIssuesCount :: Lens' AssessmentEvidenceFolder (Maybe Int)

The total number of issues that were reported directly from Security Hub, Config, or both.

assessmentEvidenceFolder_controlSetId :: Lens' AssessmentEvidenceFolder (Maybe Text)

The identifier for the control set.

assessmentEvidenceFolder_assessmentReportSelectionCount :: Lens' AssessmentEvidenceFolder (Maybe Int)

The total count of evidence included in the assessment report.

assessmentEvidenceFolder_totalEvidence :: Lens' AssessmentEvidenceFolder (Maybe Int)

The total amount of evidence in the evidence folder.

assessmentEvidenceFolder_evidenceByTypeManualCount :: Lens' AssessmentEvidenceFolder (Maybe Int)

The number of evidence that falls under the manual category. This evidence is imported manually.

assessmentEvidenceFolder_date :: Lens' AssessmentEvidenceFolder (Maybe UTCTime)

The date when the first evidence was added to the evidence folder.

assessmentEvidenceFolder_name :: Lens' AssessmentEvidenceFolder (Maybe Text)

The name of the specified evidence folder.

assessmentEvidenceFolder_evidenceByTypeUserActivityCount :: Lens' AssessmentEvidenceFolder (Maybe Int)

The number of evidence that falls under the user activity category. This evidence is collected from CloudTrail logs.

assessmentEvidenceFolder_controlId :: Lens' AssessmentEvidenceFolder (Maybe Text)

The unique identifier for the specified control.

assessmentEvidenceFolder_evidenceAwsServiceSourceCount :: Lens' AssessmentEvidenceFolder (Maybe Int)

The total number of Amazon Web Services resources assessed to generate the evidence.

assessmentEvidenceFolder_author :: Lens' AssessmentEvidenceFolder (Maybe Text)

The name of the user who created the evidence folder.

assessmentEvidenceFolder_id :: Lens' AssessmentEvidenceFolder (Maybe Text)

The identifier for the folder in which evidence is stored.

assessmentEvidenceFolder_dataSource :: Lens' AssessmentEvidenceFolder (Maybe Text)

The Amazon Web Service from which the evidence was collected.

assessmentEvidenceFolder_controlName :: Lens' AssessmentEvidenceFolder (Maybe Text)

The name of the control.

assessmentEvidenceFolder_evidenceByTypeComplianceCheckCount :: Lens' AssessmentEvidenceFolder (Maybe Int)

The number of evidence that falls under the compliance check category. This evidence is collected from Config or Security Hub.

assessmentEvidenceFolder_assessmentId :: Lens' AssessmentEvidenceFolder (Maybe Text)

The identifier for the specified assessment.

assessmentEvidenceFolder_evidenceByTypeConfigurationDataCount :: Lens' AssessmentEvidenceFolder (Maybe Int)

The number of evidence that falls under the configuration data category. This evidence is collected from configuration snapshots of other Amazon Web Services services such as Amazon EC2, Amazon S3, or IAM.

assessmentEvidenceFolder_evidenceResourcesIncludedCount :: Lens' AssessmentEvidenceFolder (Maybe Int)

The amount of evidence included in the evidence folder.