Network.AWS.AuditManager.Types.AssessmentReportMetadata

Description

Documentation

data AssessmentReportMetadata

The metadata objects associated with the specified assessment report.

See: newAssessmentReportMetadata smart constructor.

Constructors

AssessmentReportMetadata'

Fields

creationTime :: Maybe POSIX

Specifies when the assessment report was created.

status :: Maybe AssessmentReportStatus

The current status of the assessment report.

name :: Maybe Text

The name of the assessment report.

author :: Maybe Text

The name of the user who created the assessment report.

id :: Maybe Text

The unique identifier for the assessment report.

assessmentId :: Maybe Text

The unique identifier for the associated assessment.

description :: Maybe Text

The description of the specified assessment report.

assessmentName :: Maybe Text

The name of the associated assessment.

newAssessmentReportMetadata :: AssessmentReportMetadata

Create a value of AssessmentReportMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:AssessmentReportMetadata', assessmentReportMetadata_creationTime - Specifies when the assessment report was created.

$sel:status:AssessmentReportMetadata', assessmentReportMetadata_status - The current status of the assessment report.

$sel:name:AssessmentReportMetadata', assessmentReportMetadata_name - The name of the assessment report.

$sel:author:AssessmentReportMetadata', assessmentReportMetadata_author - The name of the user who created the assessment report.

$sel:id:AssessmentReportMetadata', assessmentReportMetadata_id - The unique identifier for the assessment report.

$sel:assessmentId:AssessmentReportMetadata', assessmentReportMetadata_assessmentId - The unique identifier for the associated assessment.

$sel:description:AssessmentReportMetadata', assessmentReportMetadata_description - The description of the specified assessment report.

$sel:assessmentName:AssessmentReportMetadata', assessmentReportMetadata_assessmentName - The name of the associated assessment.

assessmentReportMetadata_creationTime :: Lens' AssessmentReportMetadata (Maybe UTCTime)

Specifies when the assessment report was created.

assessmentReportMetadata_status :: Lens' AssessmentReportMetadata (Maybe AssessmentReportStatus)

The current status of the assessment report.

assessmentReportMetadata_name :: Lens' AssessmentReportMetadata (Maybe Text)

The name of the assessment report.

assessmentReportMetadata_author :: Lens' AssessmentReportMetadata (Maybe Text)

The name of the user who created the assessment report.

assessmentReportMetadata_id :: Lens' AssessmentReportMetadata (Maybe Text)

The unique identifier for the assessment report.

assessmentReportMetadata_assessmentId :: Lens' AssessmentReportMetadata (Maybe Text)

The unique identifier for the associated assessment.

assessmentReportMetadata_description :: Lens' AssessmentReportMetadata (Maybe Text)

The description of the specified assessment report.

assessmentReportMetadata_assessmentName :: Lens' AssessmentReportMetadata (Maybe Text)

The name of the associated assessment.