Network.AWS.AuditManager.Types.ChangeLog

Description

Documentation

data ChangeLog

The record of a change within Audit Manager, such as a modified assessment, a delegated control set, and so on.

See: newChangeLog smart constructor.

Constructors

ChangeLog'

Fields

objectName :: Maybe Text

The name of the changelog object.

createdAt :: Maybe POSIX

The time of creation for the changelog object.

objectType :: Maybe ObjectTypeEnum

The changelog object type, such as an assessment, control, or control set.

createdBy :: Maybe Text

The IAM user or role that performed the action.

action :: Maybe ActionEnum

The action performed.

newChangeLog :: ChangeLog

Create a value of ChangeLog with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:objectName:ChangeLog', changeLog_objectName - The name of the changelog object.

$sel:createdAt:ChangeLog', changeLog_createdAt - The time of creation for the changelog object.

$sel:objectType:ChangeLog', changeLog_objectType - The changelog object type, such as an assessment, control, or control set.

$sel:createdBy:ChangeLog', changeLog_createdBy - The IAM user or role that performed the action.

$sel:action:ChangeLog', changeLog_action - The action performed.

changeLog_objectName :: Lens' ChangeLog (Maybe Text)

The name of the changelog object.

changeLog_createdAt :: Lens' ChangeLog (Maybe UTCTime)

The time of creation for the changelog object.

changeLog_objectType :: Lens' ChangeLog (Maybe ObjectTypeEnum)

The changelog object type, such as an assessment, control, or control set.

changeLog_createdBy :: Lens' ChangeLog (Maybe Text)

The IAM user or role that performed the action.

changeLog_action :: Lens' ChangeLog (Maybe ActionEnum)

The action performed.