Network.AWS.AuditManager.Types.Delegation

Description

Documentation

data Delegation

The assignment of a control set to a delegate for review.

See: newDelegation smart constructor.

Constructors

Delegation'

Fields

roleType :: Maybe RoleType

The type of customer persona.

In CreateAssessment, roleType can only be PROCESS_OWNER.

In UpdateSettings, roleType can only be PROCESS_OWNER.

In BatchCreateDelegationByAssessment, roleType can only be RESOURCE_OWNER.

creationTime :: Maybe POSIX

Specifies when the delegation was created.

status :: Maybe DelegationStatus

The status of the delegation.

lastUpdated :: Maybe POSIX

Specifies when the delegation was last updated.

controlSetId :: Maybe Text

The identifier for the associated control set.

createdBy :: Maybe Text

The IAM user or role that created the delegation.

id :: Maybe Text

The unique identifier for the delegation.

assessmentId :: Maybe Text

The identifier for the associated assessment.

comment :: Maybe Text

The comment related to the delegation.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role.

assessmentName :: Maybe Text

The name of the associated assessment.

newDelegation :: Delegation

Create a value of Delegation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleType:Delegation', delegation_roleType - The type of customer persona.

In CreateAssessment, roleType can only be PROCESS_OWNER.

In UpdateSettings, roleType can only be PROCESS_OWNER.

In BatchCreateDelegationByAssessment, roleType can only be RESOURCE_OWNER.

$sel:creationTime:Delegation', delegation_creationTime - Specifies when the delegation was created.

$sel:status:Delegation', delegation_status - The status of the delegation.

$sel:lastUpdated:Delegation', delegation_lastUpdated - Specifies when the delegation was last updated.

$sel:controlSetId:Delegation', delegation_controlSetId - The identifier for the associated control set.

$sel:createdBy:Delegation', delegation_createdBy - The IAM user or role that created the delegation.

$sel:id:Delegation', delegation_id - The unique identifier for the delegation.

$sel:assessmentId:Delegation', delegation_assessmentId - The identifier for the associated assessment.

$sel:comment:Delegation', delegation_comment - The comment related to the delegation.

$sel:roleArn:Delegation', delegation_roleArn - The Amazon Resource Name (ARN) of the IAM role.

$sel:assessmentName:Delegation', delegation_assessmentName - The name of the associated assessment.

delegation_roleType :: Lens' Delegation (Maybe RoleType)

The type of customer persona.

In CreateAssessment, roleType can only be PROCESS_OWNER.

In UpdateSettings, roleType can only be PROCESS_OWNER.

In BatchCreateDelegationByAssessment, roleType can only be RESOURCE_OWNER.

delegation_creationTime :: Lens' Delegation (Maybe UTCTime)

Specifies when the delegation was created.

delegation_status :: Lens' Delegation (Maybe DelegationStatus)

The status of the delegation.

delegation_lastUpdated :: Lens' Delegation (Maybe UTCTime)

Specifies when the delegation was last updated.

delegation_controlSetId :: Lens' Delegation (Maybe Text)

The identifier for the associated control set.

delegation_createdBy :: Lens' Delegation (Maybe Text)

The IAM user or role that created the delegation.

delegation_id :: Lens' Delegation (Maybe Text)

The unique identifier for the delegation.

delegation_assessmentId :: Lens' Delegation (Maybe Text)

The identifier for the associated assessment.

delegation_comment :: Lens' Delegation (Maybe Text)

The comment related to the delegation.

delegation_roleArn :: Lens' Delegation (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role.

delegation_assessmentName :: Lens' Delegation (Maybe Text)

The name of the associated assessment.