Network.AWS.AuditManager.Types.DelegationMetadata

Description

Documentation

data DelegationMetadata

The metadata associated with the specified delegation.

See: newDelegationMetadata smart constructor.

Constructors

DelegationMetadata'

Fields

creationTime :: Maybe POSIX

Specifies when the delegation was created.

status :: Maybe DelegationStatus

The current status of the delgation.

controlSetName :: Maybe Text

Specifies the name of the control set delegated for review.

id :: Maybe Text

The unique identifier for the delegation.

assessmentId :: Maybe Text

The unique identifier for the specified assessment.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role.

assessmentName :: Maybe Text

The name of the associated assessment.

newDelegationMetadata :: DelegationMetadata

Create a value of DelegationMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DelegationMetadata', delegationMetadata_creationTime - Specifies when the delegation was created.

$sel:status:DelegationMetadata', delegationMetadata_status - The current status of the delgation.

$sel:controlSetName:DelegationMetadata', delegationMetadata_controlSetName - Specifies the name of the control set delegated for review.

$sel:id:DelegationMetadata', delegationMetadata_id - The unique identifier for the delegation.

$sel:assessmentId:DelegationMetadata', delegationMetadata_assessmentId - The unique identifier for the specified assessment.

$sel:roleArn:DelegationMetadata', delegationMetadata_roleArn - The Amazon Resource Name (ARN) of the IAM role.

$sel:assessmentName:DelegationMetadata', delegationMetadata_assessmentName - The name of the associated assessment.

delegationMetadata_creationTime :: Lens' DelegationMetadata (Maybe UTCTime)

Specifies when the delegation was created.

delegationMetadata_status :: Lens' DelegationMetadata (Maybe DelegationStatus)

The current status of the delgation.

delegationMetadata_controlSetName :: Lens' DelegationMetadata (Maybe Text)

Specifies the name of the control set delegated for review.

delegationMetadata_id :: Lens' DelegationMetadata (Maybe Text)

The unique identifier for the delegation.

delegationMetadata_assessmentId :: Lens' DelegationMetadata (Maybe Text)

The unique identifier for the specified assessment.

delegationMetadata_roleArn :: Lens' DelegationMetadata (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role.

delegationMetadata_assessmentName :: Lens' DelegationMetadata (Maybe Text)

The name of the associated assessment.