Network.AWS.AuditManager.Types.Evidence

Description

Documentation

data Evidence

A record that contains the information needed to demonstrate compliance with the requirements specified by a control. Examples of evidence include change activity triggered by a user, or a system configuration snapshot.

See: newEvidence smart constructor.

Constructors

Evidence'

Fields

time :: Maybe POSIX

The timestamp that represents when the evidence was collected.

assessmentReportSelection :: Maybe Text

Specifies whether the evidence is included in the assessment report.

evidenceByType :: Maybe Text

The type of automated evidence.

complianceCheck :: Maybe Text

The evaluation status for evidence that falls under the compliance check category. For evidence collected from Security Hub, a Pass or Fail result is shown. For evidence collected from Config, a Compliant or Noncompliant result is shown.

awsOrganization :: Maybe Text

The Amazon Web Services account from which the evidence is collected, and its organization path.

awsAccountId :: Maybe Text

The identifier for the specified Amazon Web Services account.

attributes :: Maybe (HashMap Text Text)

The names and values used by the evidence event, including an attribute name (such as allowUsersToChangePassword) and value (such as true or false).

evidenceAwsAccountId :: Maybe Text

The identifier for the specified Amazon Web Services account.

id :: Maybe Text

The identifier for the evidence.

dataSource :: Maybe Text

The data source from which the specified evidence was collected.

evidenceFolderId :: Maybe Text

The identifier for the folder in which the evidence is stored.

iamId :: Maybe Text

The unique identifier for the IAM user or role associated with the evidence.

eventName :: Maybe Text

The name of the specified evidence event.

resourcesIncluded :: Maybe [Resource]

The list of resources assessed to generate the evidence.

eventSource :: Maybe Text

The Amazon Web Service from which the evidence is collected.

newEvidence :: Evidence

Create a value of Evidence with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:time:Evidence', evidence_time - The timestamp that represents when the evidence was collected.

$sel:assessmentReportSelection:Evidence', evidence_assessmentReportSelection - Specifies whether the evidence is included in the assessment report.

$sel:evidenceByType:Evidence', evidence_evidenceByType - The type of automated evidence.

$sel:complianceCheck:Evidence', evidence_complianceCheck - The evaluation status for evidence that falls under the compliance check category. For evidence collected from Security Hub, a Pass or Fail result is shown. For evidence collected from Config, a Compliant or Noncompliant result is shown.

$sel:awsOrganization:Evidence', evidence_awsOrganization - The Amazon Web Services account from which the evidence is collected, and its organization path.

$sel:awsAccountId:Evidence', evidence_awsAccountId - The identifier for the specified Amazon Web Services account.

$sel:attributes:Evidence', evidence_attributes - The names and values used by the evidence event, including an attribute name (such as allowUsersToChangePassword) and value (such as true or false).

$sel:evidenceAwsAccountId:Evidence', evidence_evidenceAwsAccountId - The identifier for the specified Amazon Web Services account.

$sel:id:Evidence', evidence_id - The identifier for the evidence.

$sel:dataSource:Evidence', evidence_dataSource - The data source from which the specified evidence was collected.

$sel:evidenceFolderId:Evidence', evidence_evidenceFolderId - The identifier for the folder in which the evidence is stored.

$sel:iamId:Evidence', evidence_iamId - The unique identifier for the IAM user or role associated with the evidence.

$sel:eventName:Evidence', evidence_eventName - The name of the specified evidence event.

$sel:resourcesIncluded:Evidence', evidence_resourcesIncluded - The list of resources assessed to generate the evidence.

$sel:eventSource:Evidence', evidence_eventSource - The Amazon Web Service from which the evidence is collected.

evidence_time :: Lens' Evidence (Maybe UTCTime)

The timestamp that represents when the evidence was collected.

evidence_assessmentReportSelection :: Lens' Evidence (Maybe Text)

Specifies whether the evidence is included in the assessment report.

evidence_evidenceByType :: Lens' Evidence (Maybe Text)

The type of automated evidence.

evidence_complianceCheck :: Lens' Evidence (Maybe Text)

The evaluation status for evidence that falls under the compliance check category. For evidence collected from Security Hub, a Pass or Fail result is shown. For evidence collected from Config, a Compliant or Noncompliant result is shown.

evidence_awsOrganization :: Lens' Evidence (Maybe Text)

The Amazon Web Services account from which the evidence is collected, and its organization path.

evidence_awsAccountId :: Lens' Evidence (Maybe Text)

The identifier for the specified Amazon Web Services account.

evidence_attributes :: Lens' Evidence (Maybe (HashMap Text Text))

The names and values used by the evidence event, including an attribute name (such as allowUsersToChangePassword) and value (such as true or false).

evidence_evidenceAwsAccountId :: Lens' Evidence (Maybe Text)

The identifier for the specified Amazon Web Services account.

evidence_id :: Lens' Evidence (Maybe Text)

The identifier for the evidence.

evidence_dataSource :: Lens' Evidence (Maybe Text)

The data source from which the specified evidence was collected.

evidence_evidenceFolderId :: Lens' Evidence (Maybe Text)

The identifier for the folder in which the evidence is stored.

evidence_iamId :: Lens' Evidence (Maybe Text)

The unique identifier for the IAM user or role associated with the evidence.

evidence_eventName :: Lens' Evidence (Maybe Text)

The name of the specified evidence event.

evidence_resourcesIncluded :: Lens' Evidence (Maybe [Resource])

The list of resources assessed to generate the evidence.

evidence_eventSource :: Lens' Evidence (Maybe Text)

The Amazon Web Service from which the evidence is collected.