Network.AWS.AuditManager.UpdateAssessmentFramework

Description

Updates a custom framework in Audit Manager.

Creating a Request

data UpdateAssessmentFramework

See: newUpdateAssessmentFramework smart constructor.

Constructors

UpdateAssessmentFramework'

Fields

complianceType :: Maybe Text

The compliance type that the new custom framework supports, such as CIS or HIPAA.

description :: Maybe Text

The description of the framework that is to be updated.

frameworkId :: Text

The identifier for the specified framework.

name :: Text

The name of the framework to be updated.

controlSets :: [UpdateAssessmentFrameworkControlSet]

The control sets associated with the framework.

newUpdateAssessmentFramework

Arguments

:: Text
$sel:frameworkId:UpdateAssessmentFramework'

-> Text
$sel:name:UpdateAssessmentFramework'

-> UpdateAssessmentFramework

Create a value of UpdateAssessmentFramework with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:complianceType:UpdateAssessmentFramework', updateAssessmentFramework_complianceType - The compliance type that the new custom framework supports, such as CIS or HIPAA.

$sel:description:UpdateAssessmentFramework', updateAssessmentFramework_description - The description of the framework that is to be updated.

$sel:frameworkId:UpdateAssessmentFramework', updateAssessmentFramework_frameworkId - The identifier for the specified framework.

$sel:name:UpdateAssessmentFramework', updateAssessmentFramework_name - The name of the framework to be updated.

$sel:controlSets:UpdateAssessmentFramework', updateAssessmentFramework_controlSets - The control sets associated with the framework.

Request Lenses

updateAssessmentFramework_complianceType :: Lens' UpdateAssessmentFramework (Maybe Text)

The compliance type that the new custom framework supports, such as CIS or HIPAA.

updateAssessmentFramework_description :: Lens' UpdateAssessmentFramework (Maybe Text)

The description of the framework that is to be updated.

updateAssessmentFramework_frameworkId :: Lens' UpdateAssessmentFramework Text

The identifier for the specified framework.

updateAssessmentFramework_name :: Lens' UpdateAssessmentFramework Text

The name of the framework to be updated.

updateAssessmentFramework_controlSets :: Lens' UpdateAssessmentFramework [UpdateAssessmentFrameworkControlSet]

The control sets associated with the framework.

Destructuring the Response

data UpdateAssessmentFrameworkResponse

See: newUpdateAssessmentFrameworkResponse smart constructor.

Constructors

UpdateAssessmentFrameworkResponse'

Fields

framework :: Maybe Framework

The name of the specified framework.

httpStatus :: Int

The response's http status code.

newUpdateAssessmentFrameworkResponse

Arguments

:: Int
$sel:httpStatus:UpdateAssessmentFrameworkResponse'

-> UpdateAssessmentFrameworkResponse

Create a value of UpdateAssessmentFrameworkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:framework:UpdateAssessmentFrameworkResponse', updateAssessmentFrameworkResponse_framework - The name of the specified framework.

$sel:httpStatus:UpdateAssessmentFrameworkResponse', updateAssessmentFrameworkResponse_httpStatus - The response's http status code.

Response Lenses

updateAssessmentFrameworkResponse_framework :: Lens' UpdateAssessmentFrameworkResponse (Maybe Framework)

The name of the specified framework.

updateAssessmentFrameworkResponse_httpStatus :: Lens' UpdateAssessmentFrameworkResponse Int

The response's http status code.