Network.AWS.AuditManager.UpdateSettings

Description

Updates Audit Manager settings for the current user account.

Creating a Request

data UpdateSettings

See: newUpdateSettings smart constructor.

Constructors

UpdateSettings'

Fields

kmsKey :: Maybe Text

The KMS key details.

defaultAssessmentReportsDestination :: Maybe AssessmentReportsDestination

The default storage destination for assessment reports.

snsTopic :: Maybe Text

The Amazon Simple Notification Service (Amazon SNS) topic to which Audit Manager sends notifications.

defaultProcessOwners :: Maybe [Role]

A list of the default audit owners.

newUpdateSettings :: UpdateSettings

Create a value of UpdateSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKey:UpdateSettings', updateSettings_kmsKey - The KMS key details.

$sel:defaultAssessmentReportsDestination:UpdateSettings', updateSettings_defaultAssessmentReportsDestination - The default storage destination for assessment reports.

$sel:snsTopic:UpdateSettings', updateSettings_snsTopic - The Amazon Simple Notification Service (Amazon SNS) topic to which Audit Manager sends notifications.

$sel:defaultProcessOwners:UpdateSettings', updateSettings_defaultProcessOwners - A list of the default audit owners.

Request Lenses

updateSettings_kmsKey :: Lens' UpdateSettings (Maybe Text)

The KMS key details.

updateSettings_defaultAssessmentReportsDestination :: Lens' UpdateSettings (Maybe AssessmentReportsDestination)

The default storage destination for assessment reports.

updateSettings_snsTopic :: Lens' UpdateSettings (Maybe Text)

The Amazon Simple Notification Service (Amazon SNS) topic to which Audit Manager sends notifications.

updateSettings_defaultProcessOwners :: Lens' UpdateSettings (Maybe [Role])

A list of the default audit owners.

Destructuring the Response

data UpdateSettingsResponse

See: newUpdateSettingsResponse smart constructor.

Constructors

UpdateSettingsResponse'

Fields

settings :: Maybe Settings

The current list of settings.

httpStatus :: Int

The response's http status code.

newUpdateSettingsResponse

Arguments

:: Int
$sel:httpStatus:UpdateSettingsResponse'

-> UpdateSettingsResponse

Create a value of UpdateSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:settings:UpdateSettingsResponse', updateSettingsResponse_settings - The current list of settings.

$sel:httpStatus:UpdateSettingsResponse', updateSettingsResponse_httpStatus - The response's http status code.

Response Lenses

updateSettingsResponse_settings :: Lens' UpdateSettingsResponse (Maybe Settings)

The current list of settings.

updateSettingsResponse_httpStatus :: Lens' UpdateSettingsResponse Int

The response's http status code.