Network.AWS.Auth

Description

Documentation

data Credentials

Constructors

CredBasic ByteString ByteString
Basic credentials containing an access key and a secret key.

CredSession ByteString ByteString ByteString
Session credentials containing access key, secret key, and a security token.

CredProfile ByteString
A specific IAM Profile name to query the local instance-data for credentials.

CredEnv String String
Environment variable names to read for the access and secret keys.

CredDiscover
Attempt to read the default access and secret keys from the environment, falling back to the first available IAM profile if they are not set.

This attempts to resolve http://instance-data rather than directly retrieving http://169.254.169.254 for IAM profile information to ensure the request terminates promptly if not running on EC2.

accessKey :: String

Default access key environment variable: AWS_ACCESS_KEY

secretKey :: String

Default secret key environment variable: AWS_SECRET_KEY

credentials :: (Applicative m, MonadIO m) => Credentials -> EitherT AWSError m (IORef Auth)

defaultProfile :: (Applicative m, MonadIO m) => EitherT AWSError m ByteString

fromProfile :: (Applicative m, MonadIO m) => ByteString -> EitherT AWSError m (IORef Auth)

The IORef wrapper + timer is designed so that multiple concurrenct accesses of Auth from the AWS environment are not required to calculate expiry and sequentially queue to update it.

The forked timer ensures a singular owner and pre-emptive refresh of the temporary session credentials.