Network.AWS.AutoScaling

Description

Auto Scaling is a web service designed to automatically launch or terminate Amazon Elastic Compute Cloud (Amazon EC2) instances based on user-defined policies, schedules, and health checks.

This service is used in conjunction with Amazon CloudWatch and Elastic Load Balancing services.

Actions

CreateAutoScalingGroup

data CreateAutoScalingGroup

Creates a new Auto Scaling group with the specified name and other attributes. When the creation request is completed, the Auto Scaling group is ready to be used in other calls. Note The Auto Scaling group name must be unique within the scope of your AWS account.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_CreateAutoScalingGroup.html

Constructors

CreateAutoScalingGroup

Fields

casgAutoScalingGroupName :: !Text

The name of the Auto Scaling group.

casgAvailabilityZones :: Members AvailabilityZone

A list of Availability Zones for the Auto Scaling group. This is required unless you have specified subnets.

casgDefaultCooldown :: Maybe Integer

The amount of time, in seconds, between a successful scaling activity and the succeeding scaling activity.

casgDesiredCapacity :: Maybe Integer

The number of Amazon EC2 instances that should be running in the group. The desired capacity must be greater than or equal to the minimum size and less than or equal to the maximum size specified for the Auto Scaling group.

casgHealthCheckGracePeriod :: Maybe Integer

Length of time in seconds after a new Amazon EC2 instance comes into service that Auto Scaling starts checking its health. During this time any health check failure for the that instance is ignored.

casgHealthCheckType :: Maybe Text

The service you want the health checks from, Amazon EC2 or Elastic Load Balancer. Valid values are EC2 or ELB.

casgLaunchConfigurationName :: !Text

The name of an existing launch configuration to use to launch new instances.

casgLoadBalancerNames :: Members Text

A list of existing Elastic Load Balancing load balancers to use. The load balancers must be associated with the AWS account.

casgMaxSize :: !Integer

The maximum size of the Auto Scaling group.

casgMinSize :: !Integer

The minimum size of the Auto Scaling group.

casgPlacementGroup :: Maybe Text

Physical location of an existing cluster placement group into which you want to launch your instances. For information about cluster placement group, see Using Cluster Instances

casgTags :: Members Tag

The tag to be created or updated. Each tag should be defined by its resource type, resource ID, key, value, and a propagate flag. Valid values: key=value, value=value, propagate=true or false. Value and propagate are optional parameters.

casgTerminationPolicies :: Members Text

A standalone termination policy or a list of termination policies used to select the instance to terminate. The policies are executed in the order that they are listed.

casgVPCZoneIdentifier :: Maybe Text

A comma-separated list of subnet identifiers of Amazon Virtual Private Clouds (Amazon VPCs).

data CreateAutoScalingGroupResponse

Constructors

CreateAutoScalingGroupResponse

Fields

casgrResponseMetadata :: !ResponseMetadata

CreateLaunchConfiguration

data CreateLaunchConfiguration

Creates a new launch configuration. The launch configuration name must be unique within the scope of the client's AWS account.

The maximum limit of launch configurations, which by default is 100, must not yet have been met; otherwise, the call will fail.

When created, the new launch configuration is available for immediate use.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_CreateLaunchConfiguration.html

Constructors

CreateLaunchConfiguration

Fields

clcBlockDeviceMappings :: Members BlockDeviceMapping

A list of mappings that specify how block devices are exposed to the instance. Each mapping is made up of a VirtualName, a DeviceName, and an ebs data structure that contains information about the associated Elastic Block Storage volume. For more information about Amazon EC2 BlockDeviceMappings, go to Block Device Mapping in the Amazon EC2 product documentation.

clcEbsOptimized :: Maybe Bool

Whether the instance is optimized for EBS I/O. The optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization is not available with all instance types. Additional usage charges apply when using an EBS Optimized instance. By default the instance is not optimized for EBS I/O. For information about EBS-optimized instances, go to EBS-Optimized Instances in the Amazon Elastic Compute Cloud User Guide.

clcIamInstanceProfile :: Maybe Text

The name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance.

clcImageId :: !Text

Unique ID of the Amazon Machine Image (AMI) you want to use to launch your EC2 instances. For information about finding Amazon EC2 AMIs, see Finding a Suitable AMI in the Amazon Elastic Compute Cloud User Guide.

clcInstanceMonitoring :: Maybe InstanceMonitoring

Enables detailed monitoring if it is disabled. Detailed monitoring is enabled by default.

clcInstanceType :: !InstanceType

The instance type of the Amazon EC2 instance. For information about available Amazon EC2 instance types, see Available Instance Types in the Amazon Elastic Cloud Compute User Guide.

clcKernelId :: Maybe Text

The ID of the kernel associated with the Amazon EC2 AMI.

clcKeyName :: Maybe Text

The name of the Amazon EC2 key pair. For more information, see Getting a Key Pair in the Amazon Elastic Compute Cloud User Guide.

clcLaunchConfigurationName :: !Text

The name of the launch configuration to create.

clcRamdiskId :: Maybe Text

The ID of the RAM disk associated with the Amazon EC2 AMI.

clcSecurityGroups :: Members Text

The security groups with which to associate Amazon EC2 or Amazon VPC instances.

clcSpotPrice :: Maybe Text

The maximum hourly price to be paid for any Spot Instance launched to fulfill the request. Spot Instances are launched when the price you specify exceeds the current Spot market price. For more information on launching Spot Instances, see Using Auto Scaling to Launch Spot Instances in the Auto Scaling Developer Guide.

clcUserData :: Maybe Text

The user data to make available to the launched Amazon EC2 instances. For more information about Amazon EC2 user data, see User Data Retrieval in the Amazon Elastic Compute Cloud User Guide.

data CreateLaunchConfigurationResponse

Constructors

CreateLaunchConfigurationResponse

Fields

clcrResponseMetadata :: !ResponseMetadata

CreateOrUpdateTags

data CreateOrUpdateTags

Creates new tags or updates existing tags for an Auto Scaling group. Note A tag's definition is composed of a resource ID, resource type, key and value, and the propagate flag. Value and the propagate flag are optional parameters. See the Request Parameters for more information. For information on creating tags for your Auto Scaling group, see Tag Your Auto Scaling Groups and Amazon EC2 Instances.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_CreateOrUpdateTags.html

Constructors

CreateOrUpdateTags

Fields

coutTags :: Members Tag

The tag to be created or updated. Each tag should be defined by its resource type, resource ID, key, value, and a propagate flag. The resource type and resource ID identify the type and name of resource for which the tag is created. Currently, auto-scaling-group is the only supported resource type. The valid value for the resource ID is groupname.

data CreateOrUpdateTagsResponse

Constructors

CreateOrUpdateTagsResponse

Fields

coutrResponseMetadata :: !ResponseMetadata

DeleteAutoScalingGroup

data DeleteAutoScalingGroup

Deletes the specified Auto Scaling group if the group has no instances and no scaling activities in progress. Note To remove all instances before calling DeleteAutoScalingGroup, you can call UpdateAutoScalingGroup to set the minimum and maximum size of the AutoScalingGroup to zero.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DeleteAutoScalingGroup.html

Constructors

DeleteAutoScalingGroup

Fields

dasgAutoScalingGroupName :: !Text

The name of the Auto Scaling group to delete.

dasgForceDelete :: Maybe Bool

Starting with API version 2011-01-01, specifies that the Auto Scaling group will be deleted along with all instances associated with the group, without waiting for all instances to be terminated.

data DeleteAutoScalingGroupResponse

Constructors

DeleteAutoScalingGroupResponse

Fields

dasgrResponseMetadata :: !ResponseMetadata

DeleteLaunchConfiguration

data DeleteLaunchConfiguration

Deletes the specified LaunchConfiguration. The specified launch configuration must not be attached to an Auto Scaling group. When this call completes, the launch configuration is no longer available for use.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DeleteLaunchConfiguration.html

Constructors

DeleteLaunchConfiguration

Fields

dlcLaunchConfigurationName :: !Text

The name of the launch configuration.

data DeleteLaunchConfigurationResponse

Constructors

DeleteLaunchConfigurationResponse

Fields

dlcrResponseMetadata :: !ResponseMetadata

DeleteNotificationConfiguration

data DeleteNotificationConfiguration

Deletes notifications created by PutNotificationConfiguration.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DeleteNotificationConfiguration.html

Constructors

DeleteNotificationConfiguration

Fields

dncAutoScalingGroupName :: !Text

The name of the Auto Scaling group.

dncTopicARN :: !Text

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

data DeleteNotificationConfigurationResponse

Constructors

DeleteNotificationConfigurationResponse

DeletePolicy

data DeletePolicy

Deletes a policy created by PutScalingPolicy.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DeletePolicy.html

Constructors

DeletePolicy

Fields

dpAutoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

dpPolicyName :: !Text

The name or PolicyARN of the policy you want to delete.

data DeletePolicyResponse

Constructors

DeletePolicyResponse

DeleteScheduledAction

data DeleteScheduledAction

Deletes a scheduled action previously created using the PutScheduledUpdateGroupAction.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DeleteScheduledAction.html

Constructors

DeleteScheduledAction

Fields

dsaAutoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

dsaScheduledActionName :: !Text

The name of the action you want to delete.

data DeleteScheduledActionResponse

Constructors

DeleteScheduledActionResponse

DeleteTags

data DeleteTags

Removes the specified tags or a set of tags from a set of resources.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DeleteTags.html

Constructors

DeleteTags

Fields

dtTags :: Members Tag

Each tag should be defined by its resource type, resource ID, key, value, and a propagate flag. Valid values are: Resource type = auto-scaling-group, Resource ID = AutoScalingGroupName, key=value, value=value, propagate=true or false.

data DeleteTagsResponse

Constructors

DeleteTagsResponse

DescribeAdjustmentTypes

data DescribeAdjustmentTypes

Returns policy adjustment types for use in the PutScalingPolicy action.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeAdjustmentTypes.html

Constructors

DescribeAdjustmentTypes

data DescribeAdjustmentTypesResponse

Constructors

DescribeAdjustmentTypesResponse

Fields

datrDescribeAdjustmentTypesResult :: !DescribeAdjustmentTypesResult

DescribeAutoScalingGroups

data DescribeAutoScalingGroups

Returns a full description of each Auto Scaling group in the given list. This includes all Amazon EC2 instances that are members of the group. If a list of names is not provided, the service returns the full details of all Auto Scaling groups. This action supports pagination by returning a token if there are more pages to retrieve. To get the next page, call this action again with the returned token as the NextToken parameter.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeAutoScalingGroups.html

Constructors

DescribeAutoScalingGroups

Fields

dasgAutoScalingGroupNames :: Members Text

A list of Auto Scaling group names.

dasgMaxRecords :: Maybe Integer

The maximum number of records to return.

dasgNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

data DescribeAutoScalingGroupsResponse

Constructors

DescribeAutoScalingGroupsResponse

Fields

dashrDescribeAutoScalingGroupsResult :: !DescribeAutoScalingGroupsResult

DescribeAutoScalingInstances

data DescribeAutoScalingInstances

Returns a description of each Auto Scaling instance in the InstanceIds list. If a list is not provided, the service returns the full details of all instances up to a maximum of 50. By default, the service returns a list of 20 items. This action supports pagination by returning a token if there are more pages to retrieve. To get the next page, call this action again with the returned token as the NextToken parameter.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeAutoScalingInstances.html

Constructors

DescribeAutoScalingInstances

Fields

dasiInstanceIds :: Members Text

The list of Auto Scaling instances to describe. If this list is omitted, all auto scaling instances are described. The list of requested instances cannot contain more than 50 items. If unknown instances are requested, they are ignored with no error.

dasiMaxRecords :: Maybe Integer

The maximum number of Auto Scaling instances to be described with each call.

dasiNextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available.

data DescribeAutoScalingInstancesResponse

Constructors

DescribeAutoScalingInstancesResponse

Fields

dasirDescribeAutoScalingInstancesResult :: !DescribeAutoScalingInstancesResult

DescribeAutoScalingNotificationTypes

data DescribeAutoScalingNotificationTypes

Returns a list of all notification types that are supported by Auto Scaling.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeAutoScalingNotificationTypes.html

Constructors

DescribeAutoScalingNotificationTypes

Fields

dasntAutoScalingNotificationTypes :: !Text

Returns a list of all notification types supported by Auto Scaling. They are:

data DescribeAutoScalingNotificationTypesResponse

Constructors

DescribeAutoScalingNotificationTypesResponse

Fields

dasntrDescribeAutoScalingNotificationTypesResult :: !DescribeAutoScalingNotificationTypesResult

DescribeLaunchConfigurations

data DescribeLaunchConfigurations

Returns a full description of the launch configurations, or the specified launch configurations, if they exist. If no name is specified, then the full details of all launch configurations are returned.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeLaunchConfigurations.html

Constructors

DescribeLaunchConfigurations

Fields

dlcLaunchConfigurationNames :: Members Text

A list of launch configuration names.

dlcMaxRecords :: Maybe Integer

The maximum number of launch configurations. The default is 100.

dlcNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

data DescribeLaunchConfigurationsResponse

Constructors

DescribeLaunchConfigurationsResponse

Fields

dldrDescribeLaunchConfigurationsResult :: !DescribeLaunchConfigurationsResult

DescribeMetricCollectionTypes

data DescribeMetricCollectionTypes

Returns a list of metrics and a corresponding list of granularities for each metric.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeMetricCollectionTypes.html

Constructors

DescribeMetricCollectionTypes

data DescribeMetricCollectionTypesResponse

Constructors

DescribeMetricCollectionTypesResponse

Fields

dmctDescribeMetricCollectionTypesResult :: !DescribeMetricCollectionTypesResult

DescribeNotificationConfigurations

data DescribeNotificationConfigurations

Returns a list of notification actions associated with Auto Scaling groups for specified events.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeNotificationConfigurations.html

Constructors

DescribeNotificationConfigurations

Fields

dncAutoScalingGroupNames :: Members Text

The name of the Auto Scaling group.

dncMaxRecords :: Maybe Integer

Maximum number of records to be returned.

dncNextToken :: Maybe Text

A string that is used to mark the start of the next batch of returned results for pagination.

data DescribeNotificationConfigurationsResponse

Constructors

DescribeNotificationConfigurationsResponse

Fields

dndrDescribeNotificationConfigurationsResult :: !DescribeNotificationConfigurationsResult

DescribePolicies

data DescribePolicies

Returns descriptions of what each policy does. This action supports pagination. If the response includes a token, there are more records available. To get the additional records, repeat the request with the response token as the NextToken parameter.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribePolicies.html

Constructors

DescribePolicies

Fields

dqAutoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

dqMaxRecords :: Maybe Integer

The maximum number of policies that will be described with each call.

dqNextToken :: Maybe Text

A string that is used to mark the start of the next batch of returned results for pagination.

dqPolicyNames :: Members Text

A list of policy names or policy ARNs to be described. If this list is omitted, all policy names are described. If an auto scaling group name is provided, the results are limited to that group. The list of requested policy names cannot contain more than 50 items. If unknown policy names are requested, they are ignored with no error.

data DescribePoliciesResponse

Constructors

DescribePoliciesResponse

Fields

dqrDescribePoliciesResult :: !DescribePoliciesResult

DescribeScalingActivities

data DescribeScalingActivities

Returns the scaling activities for the specified Auto Scaling group. If the specified ActivityIds list is empty, all the activities from the past six weeks are returned. Activities are sorted by completion time. Activities still in progress appear first on the list. This action supports pagination. If the response includes a token, there are more records available. To get the additional records, repeat the request with the response token as the NextToken parameter.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeScalingActivities.html

Constructors

DescribeScalingActivities

Fields

dsbActivityIds :: Members Text

A list containing the activity IDs of the desired scaling activities. If this list is omitted, all activities are described. If an AutoScalingGroupName is provided, the results are limited to that group. The list of requested activities cannot contain more than 50 items. If unknown activities are requested, they are ignored with no error.

dsbAutoScalingGroupName :: Maybe Text

The name of the AutoScalingGroup.

dsbMaxRecords :: Maybe Integer

The maximum number of scaling activities to return.

dsbNextToken :: Maybe Text

A string that marks the start of the next batch of returned results for pagination.

data DescribeScalingActivitiesResponse

Constructors

DescribeScalingActivitiesResponse

Fields

dsbrDescribeScalingActivitiesResult :: !DescribeScalingActivitiesResult

DescribeScalingProcessTypes

data DescribeScalingProcessTypes

Returns scaling process types for use in the ResumeProcesses and SuspendProcesses actions.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeScalingProcessTypes.html

Constructors

DescribeScalingProcessTypes

Fields

dsptProcesses :: !ProcessType

A list of ProcessType names.

data DescribeScalingProcessTypesResponse

Constructors

DescribeScalingProcessTypesResponse

Fields

dsptrDescribeScalingProcessTypesResult :: !DescribeScalingProcessTypesResult

DescribeScheduledActions

data DescribeScheduledActions

Lists all the actions scheduled for your Auto Scaling group that haven't been executed. To see a list of actions already executed, see the activity record returned in DescribeScalingActivities.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeScheduledActions.html

Constructors

DescribeScheduledActions

Fields

dscAutoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

dscEndTime :: Maybe UTCTime

The latest scheduled start time to return. If scheduled action names are provided, this field is ignored.

dscMaxRecords :: Maybe Integer

The maximum number of scheduled actions to return.

dscNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

dscScheduledActionNames :: Members Text

A list of scheduled actions to be described. If this list is omitted, all scheduled actions are described. The list of requested scheduled actions cannot contain more than 50 items. If an auto scaling group name is provided, the results are limited to that group. If unknown scheduled actions are requested, they are ignored with no error.

dscStartTime :: Maybe UTCTime

The earliest scheduled start time to return. If scheduled action names are provided, this field will be ignored.

data DescribeScheduledActionsResponse

Constructors

DescribeScheduledActionsResponse

Fields

dscrDescribeScheduledActionsResult :: !DescribeScheduledActionsResult

DescribeTags

data DescribeTags

Lists the Auto Scaling group tags. You can use filters to limit results when describing tags. For example, you can query for tags of a particular Auto Scaling group. You can specify multiple values for a filter. A tag must match at least one of the specified values for it to be included in the results. You can also specify multiple filters. The result includes information for a particular tag only if it matches all your filters. If there's no match, no special message is returned.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeTags.html

Constructors

DescribeTags

Fields

dtFilters :: Members Filter

The value of the filter type used to identify the tags to be returned. For example, you can filter so that tags are returned according to Auto Scaling group, the key and value, or whether the new tag will be applied to instances launched after the tag is created (PropagateAtLaunch).

dtMaxRecords :: Maybe Integer

The maximum number of records to return.

dtNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

data DescribeTagsResponse

Constructors

DescribeTagsResponse

Fields

dtrDescribeTagsResult :: !DescribeTagsResult

DescribeTerminationPolicyTypes

data DescribeTerminationPolicyTypes

Returns a list of all termination policies supported by Auto Scaling.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeTerminationPolicyTypes.html

Constructors

DescribeTerminationPolicyTypes

Fields

dtptTerminationPolicyTypes :: !Text

Termination policies supported by Auto Scaling. They are: OldestInstance, OldestLaunchConfiguration, NewestInstance, ClosestToNextInstanceHour, Default

data DescribeTerminationPolicyTypesResponse

Constructors

DescribeTerminationPolicyTypesResponse

Fields

dtptrDescribeTerminationPolicyTypesResult :: !DescribeTerminationPolicyTypesResult

DisableMetricsCollection

data DisableMetricsCollection

Disables monitoring of group metrics for the Auto Scaling group specified in AutoScalingGroupName. You can specify the list of affected metrics with the Metrics parameter.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DisableMetricsCollection.html

Constructors

DisableMetricsCollection

Fields

dmcAutoScalingGroupName :: !Text

The name or ARN of the Auto Scaling Group.

dmcMetrics :: Members Text

The list of metrics to disable. If no metrics are specified, all metrics are disabled. The following metrics are supported:

data DisableMetricsCollectionResponse

Constructors

DisableMetricsCollectionResponse

Fields

dmcrResponseMetadata :: !ResponseMetadata

EnableMetricsCollection

data EnableMetricsCollection

Enables monitoring of group metrics for the Auto Scaling group specified in AutoScalingGroupName. You can specify the list of enabled metrics with the Metrics parameter. Auto Scaling metrics collection can be turned on only if the InstanceMonitoring flag, in the Auto Scaling group's launch configuration, is set to True.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_EnableMetricsCollection.html

Constructors

EnableMetricsCollection

Fields

emcAutoScalingGroupName :: !Text

The name or ARN of the Auto Scaling group.

emcGranularity :: !Text

The granularity to associate with the metrics to collect. Currently, the only legal granularity is 1Minute.

emcMetrics :: Members Text

The list of metrics to collect. If no metrics are specified, all metrics are enabled. The following metrics are supported:

data EnableMetricsCollectionResponse

Constructors

EnableMetricsCollectionResponse

Fields

emcrResponseMetadata :: !ResponseMetadata

ExecutePolicy

data ExecutePolicy

Executes the specified policy.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_ExecutePolicy.html

Constructors

ExecutePolicy

Fields

epAutoScalingGroupName :: Maybe Text

The name or the Amazon Resource Name (ARN) of the Auto Scaling group.

epHonorCooldown :: Maybe Bool

Set to True if you want Auto Scaling to wait for the cooldown period associated with the Auto Scaling group to complete before executing the policy.

epPolicyName :: !Text

The name or ARN of the policy you want to run.

data ExecutePolicyResponse

Constructors

ExecutePolicyResponse

Fields

eprResponseMetadata :: !ResponseMetadata

PutNotificationConfiguration

data PutNotificationConfiguration

Configures an Auto Scaling group to send notifications when specified events take place. Subscribers to this topic can have messages for events delivered to an endpoint such as a web server or email address. For more information see Get Email Notifications When Your Auto Scaling Group Changes A new PutNotificationConfiguration overwrites an existing configuration.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_PutNotificationConfiguration.html

Constructors

PutNotificationConfiguration

Fields

pncAutoScalingGroupName :: !Text

The name of the Auto Scaling group.

pncNotificationTypes :: Members Text

The type of event that will cause the notification to be sent. For details about notification types supported by Auto Scaling, see DescribeAutoScalingNotificationTypes.

pncTopicARN :: !Text

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

data PutNotificationConfigurationResponse

Constructors

PutNotificationConfigurationResponse

Fields

pncrResponseMetadata :: !ResponseMetadata

PutScalingPolicy

data PutScalingPolicy

Creates or updates a policy for an Auto Scaling group. To update an existing policy, use the existing policy name and set the parameter(s) you want to change. Any existing parameter not changed in an update to an existing policy is not changed in this update request.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_PutScalingPolicy.html

Constructors

PutScalingPolicy

Fields

pspAdjustmentType :: !Text

Specifies whether the ScalingAdjustment is an absolute number or a percentage of the current capacity. Valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

pspAutoScalingGroupName :: !Text

The name or ARN of the Auto Scaling group.

pspCooldown :: Maybe Integer

The amount of time, in seconds, after a scaling activity completes and before the next scaling acitvity can start.

pspMinAdjustmentStep :: Maybe Integer

Used with AdjustmentType with the value PercentChangeInCapacity, the scaling policy changes the DesiredCapacity of the Auto Scaling group by at least the number of instances specified in the value.

pspPolicyName :: !Text

The name of the policy you want to create or update.

pspScalingAdjustment :: !Integer

The number of instances by which to scale. AdjustmentType determines the interpretation of this number (e.g., as an absolute number or as a percentage of the existing Auto Scaling group size). A positive increment adds to the current capacity and a negative value removes from the current capacity.

data PutScalingPolicyResponse

Constructors

PutScalingPolicyResponse

Fields

psprPutScalingPolicyResult :: !PutScalingPolicyResult

PutScheduledUpdateGroupAction

data PutScheduledUpdateGroupAction

Creates or updates a scheduled scaling action for an Auto Scaling group. When updating a scheduled scaling action, if you leave a parameter unspecified, the corresponding value remains unchanged in the affected Auto Scaling group. For information on creating or updating a scheduled action for your Auto Scaling group, see Scale Based on a Schedule. Note Auto Scaling supports the date and time expressed in YYYY-MM-DDThh:mm:ssZ format in UTC/GMT only.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_PutScheduledUpdateGroupAction.html

Constructors

PutScheduledUpdateGroupAction

Fields

psugaAutoScalingGroupName :: !Text

The name or ARN of the Auto Scaling group.

psugaDesiredCapacity :: Maybe Integer

The number of Amazon EC2 instances that should be running in the group.

psugaEndTime :: Maybe UTCTime

The time for this action to end.

psugaMaxSize :: Maybe Integer

The maximum size for the Auto Scaling group.

psugaMinSize :: Maybe Integer

The minimum size for the new Auto Scaling group.

psugaRecurrence :: Maybe Text

The time when recurring future actions will start. Start time is specified by the user following the Unix cron syntax format. For information about cron syntax, go to Wikipedia, The Free Encyclopedia.

psugaScheduledActionName :: !Text

The name of this scaling action.

psugaStartTime :: Maybe UTCTime

The time for this action to start, as in --start-time 2010-06-01T00:00:00Z.

psugaTime :: Maybe UTCTime

Time is deprecated.

data PutScheduledUpdateGroupActionResponse

Constructors

PutScheduledUpdateGroupActionResponse

Fields

psugarResponseMetadata :: !ResponseMetadata

ResumeProcesses

data ResumeProcesses

Resumes all suspended Auto Scaling processes for an Auto Scaling group. For information on suspending and resuming Auto Scaling process, see Suspend and Resume Auto Scaling Process.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_ResumeProcesses.html

Constructors

ResumeProcesses

Fields

rpAutoScalingGroupName :: !Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

rpScalingProcesses :: Members Text

The processes that you want to suspend or resume, which can include one or more of the following:

data ResumeProcessesResponse

Constructors

ResumeProcessesResponse

Fields

rprResponseMetadata :: !ResponseMetadata

SetDesiredCapacity

data SetDesiredCapacity

Sets the desired size of the specified AutoScalingGroup.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_SetDesiredCapacity.html

Constructors

SetDesiredCapacity

Fields

sdcAutoScalingGroupName :: !Text

The name of the Auto Scaling group.

sdcDesiredCapacity :: !Integer

The new capacity setting for the Auto Scaling group.

sdcHonorCooldown :: Maybe Bool

By default, SetDesiredCapacity overrides any cooldown period associated with the Auto Scaling group. Set to True if you want Auto Scaling to wait for the cooldown period associated with the Auto Scaling group to complete before initiating a scaling activity to set your Auto Scaling group to the new capacity setting.

data SetDesiredCapacityResponse

Constructors

SetDesiredCapacityResponse

Fields

sdcrResponseMetadata :: !ResponseMetadata

SetInstanceHealth

data SetInstanceHealth

Sets the health status of a specified instance that belongs to any of your Auto Scaling groups. For more information, see Configure Health Checks for Your Auto Scaling group.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_SetInstanceHealth.html

Constructors

SetInstanceHealth

Fields

sihHealthStatus :: !Text

The health status of the instance. Set to Healthy if you want the instance to remain in service. Set to Unhealthy if you want the instance to be out of service. Auto Scaling will terminate and replace the unhealthy instance.

sihInstanceId :: !Text

The identifier of the Amazon EC2 instance.

sihShouldRespectGracePeriod :: Maybe Bool

If the Auto Scaling group of the specified instance has a HealthCheckGracePeriod specified for the group, by default, this call will respect the grace period. Set this to False, if you do not want the call to respect the grace period associated with the group.

data SetInstanceHealthResponse

Constructors

SetInstanceHealthResponse

Fields

sihrResponseMetadata :: !ResponseMetadata

SuspendProcesses

data SuspendProcesses

Suspends Auto Scaling processes for an Auto Scaling group. To suspend specific process types, specify them by name with the ScalingProcesses.member.N parameter. To suspend all process types, omit the ScalingProcesses.member.N parameter. Important Suspending either of the two primary process types, Launch or Terminate, can prevent other process types from functioning properly. To resume processes that have been suspended, use ResumeProcesses For more information on suspending and resuming Auto Scaling process, see Suspend and Resume Auto Scaling Process.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_SuspendProcesses.html

Constructors

SuspendProcesses

Fields

spAutoScalingGroupName :: !Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

spScalingProcesses :: Members Text

The processes that you want to suspend or resume, which can include one or more of the following:

data SuspendProcessesResponse

Constructors

SuspendProcessesResponse

Fields

sprResponseMetadata :: !ResponseMetadata

TerminateInstanceInAutoScalingGroup

data TerminateInstanceInAutoScalingGroup

Terminates the specified instance. Optionally, the desired group size can be adjusted. Note This call simply registers a termination request. The termination of the instance cannot happen immediately.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_TerminateInstanceInAutoScalingGroup.html

Constructors

TerminateInstanceInAutoScalingGroup

Fields

tiiasgInstanceId :: !Text

The ID of the Amazon EC2 instance to be terminated.

tiiasgShouldDecrementDesiredCapacity :: !Bool

Specifies whether (true) or not (false) terminating this instance should also decrement the size of the AutoScalingGroup.

data TerminateInstanceInAutoScalingGroupResponse

Constructors

TerminateInstanceInAutoScalingGroupResponse

Fields

tiiasgrResponseMetadata :: !ResponseMetadata

tiiasgrTerminateInstanceInAutoScalingGroupResult :: !TerminateInstanceInAutoScalingGroupResult

UpdateAutoScalingGroup

data UpdateAutoScalingGroup

Updates the configuration for the specified AutoScalingGroup.

Note: To update an Auto Scaling group with a launch configuration that has the InstanceMonitoring flag set to False, you must first ensure that collection of group metrics is disabled. Otherwise, calls to UpdateAutoScalingGroup will fail. If you have previously enabled group metrics collection, you can disable collection of all group metrics by calling DisableMetricsCollection. The new settings are registered upon the completion of this call. Any launch configuration settings take effect on any triggers after this call returns.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_UpdateAutoScalingGroup.html

Constructors

UpdateAutoScalingGroup

Fields

uasgAutoScalingGroupName :: !Text

The name of the Auto Scaling group.

uasgAvailabilityZones :: Members AvailabilityZone

Availability Zones for the group.

uasgDefaultCooldown :: Maybe Integer

The amount of time, in seconds, after a scaling activity completes before any further scaling activities can start. For more information, see Cooldown Period.

uasgDesiredCapacity :: Maybe Integer

The desired capacity for the Auto Scaling group.

uasgHealthCheckGracePeriod :: Maybe Integer

The length of time that Auto Scaling waits before checking an instance's health status. The grace period begins when an instance comes into service.

uasgHealthCheckType :: Maybe Text

The type of health check for the instances in the Auto Scaling group. The health check type can either be EC2 for Amazon EC2 or ELB for Elastic Load Balancing.

uasgLaunchConfigurationName :: Maybe Text

The name of the launch configuration.

uasgMaxSize :: Maybe Integer

The maximum size of the Auto Scaling group.

uasgMinSize :: Maybe Integer

The minimum size of the Auto Scaling group.

uasgPlacementGroup :: Maybe Text

The name of the cluster placement group, if applicable. For more information, go to Using Cluster Instances in the Amazon EC2 User Guide.

uasgTerminationPolicies :: Members Text

A standalone termination policy or a list of termination policies used to select the instance to terminate. The policies are executed in the order that they are listed.

uasgVPCZoneIdentifier :: Maybe Text

The subnet identifier for the Amazon VPC connection, if applicable. You can specify several subnets in a comma-separated list.

data UpdateAutoScalingGroupResponse

Constructors

UpdateAutoScalingGroupResponse

Fields

uasgrResponseMetadata :: !ResponseMetadata

Data Types

autoScaling :: Service

Currently supported version of the AutoScaling service.

data ErrorType

Constructors

Receiver

Sender

data AutoScalingError

Constructors

AutoScalingError

Fields

aseType :: !ErrorType

aseCode :: !Text

aseMessage :: !Text

data AutoScalingErrorResponse

Constructors

AutoScalingErrorResponse

Fields

aserError :: !AutoScalingError

aserRequestId :: !Text

data ResponseMetadata

Constructors

ResponseMetadata

Fields

rmRequestId :: !Text

newtype ResourceName

Constructors

ResourceName Text

data Activity

A scaling Activity is a long-running process that represents a change to your AutoScalingGroup, such as changing the size of the group. It can also be a process to replace an instance, or a process to perform any other long-running operations supported by the API.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_Activity.html

Constructors

Activity

Fields

aActivityId :: !Text

Specifies the ID of the activity.

aAutoScalingGroupName :: !Text

The name of the Auto Scaling group.

aCause :: !Text

Contains the reason the activity was begun.

aDescription :: Maybe Text

Contains a friendly, more verbose description of the scaling activity.

aDetails :: Maybe Text

Contains details of the scaling activity.

aEndTime :: Maybe UTCTime

Provides the end time of this activity.

aProgress :: Maybe Integer

Specifies a value between 0 and 100 that indicates the progress of the activity.

aStartTime :: !UTCTime

Provides the start time of this activity.

aStatusCode :: !Text

Contains the current status of the activity.

aStatusMessage :: Maybe Text

Contains a friendly, more verbose description of the activity status.

data AdjustmentType

Specifies whether the PutScalingPolicy ScalingAdjustment parameter is an absolute number or a percentage of the current capacity.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_AdjustmentType.html

Constructors

ChangeInCapacity

ExactCapacity

PercentChangeInCapacity

data Alarm

The Alarm data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_Alarm.html

Constructors

Alarm

Fields

aAlarmARN :: Maybe Text

The Amazon Resource Name (ARN) of the alarm.

aAlarmName :: Maybe Text

The name of the alarm.

data AutoScalingGroup

The AutoScalingGroup data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_AutoScalingGroup.html

Constructors

AutoScalingGroup

Fields

asgAutoScalingGroupARN :: Maybe Text

The Amazon Resource Name (ARN) of the Auto Scaling group.

asgAutoScalingGroupName :: !Text

Specifies the name of the group.

asgAvailabilityZones :: [AvailabilityZone]

Contains a list of Availability Zones for the group.

asgCreatedTime :: !UTCTime

Specifies the date and time the Auto Scaling group was created.

asgDefaultCooldown :: !Integer

The number of seconds after a scaling activity completes before any further scaling activities can start.

asgDesiredCapacity :: !Integer

Specifies the desired capacity for the Auto Scaling group.

asgEnabledMetrics :: Members EnabledMetric

A list of metrics enabled for this Auto Scaling group.

asgHealthCheckGracePeriod :: Maybe Integer

The length of time that Auto Scaling waits before checking an instance's health status. The grace period begins when an instance comes into service.

asgHealthCheckType :: !Text

The service of interest for the health status check, either EC2 for Amazon EC2 or ELB for Elastic Load Balancing.

asgInstances :: Members Instance

Provides a summary list of Amazon EC2 instances.

asgLaunchConfigurationName :: Maybe Text

Specifies the name of the associated LaunchConfiguration.

asgLoadBalancerNames :: [Text]

A list of load balancers associated with this Auto Scaling group.

asgMaxSize :: !Integer

Contains the maximum size of the Auto Scaling group.

asgMinSize :: !Integer

Contains the minimum size of the Auto Scaling group.

asgPlacementGroup :: Maybe Text

The name of the cluster placement group, if applicable.

asgStatus :: Maybe Text

The current state of the Auto Scaling group when a DeleteAutoScalingGroup action is in progress.

asgSuspendedProcesses :: Members SuspendedProcess

Suspended processes associated with this Auto Scaling group.

asgTags :: Members Tag

A list of tags for the Auto Scaling group.

asgTerminationPolicies :: [Text]

A standalone termination policy or a list of termination policies for this Auto Scaling group.

asgVPCZoneIdentifier :: Maybe Text

The subnet identifier for the Amazon VPC connection, if applicable. You can specify several subnets in a comma-separated list.

data AutoScalingInstanceDetails

The AutoScalingInstanceDetails data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_AutoScalingInstanceDetails.html

Constructors

AutoScalingInstanceDetails

Fields

asidAutoScalingGroupName :: !Text

The name of the Auto Scaling group associated with this instance.

asidAvailabilityZone :: !AvailabilityZone

The Availability Zone in which this instance resides.

asidHealthStatus :: !Text

The health status of this instance. Healthy means that the instance is healthy and should remain in service. Unhealthy means that the instance is unhealthy. Auto Scaling should terminate and replace it.

asidInstanceId :: !Text

The instance ID of the Amazon EC2 instance.

asidLaunchConfigurationName :: !Text

The launch configuration associated with this instance.

asidLifecycleState :: !Text

The life cycle state of this instance. for more information, see Instance Lifecycle State in the Auto Scaling Developer Guide.

data BlockDeviceMapping

The BlockDeviceMapping data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_BlockDeviceMapping.html

Constructors

BlockDeviceMapping

Fields

bdmDeviceName :: !Text

The name of the device within Amazon EC2.

bdmEbs :: Maybe Ebs

The Elastic Block Storage volume information.

bdmVirtualName :: Maybe Text

The virtual name associated with the device.

data DescribeAdjustmentTypesResult

The output of the DescribeAdjustmentTypes action.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeAdjustmentTypesResult.html

Constructors

DescribeAdjustmentTypesResult

Fields

datrAdjustmentTypes :: Members AdjustmentType

A list of specific policy adjustment types.

data DescribeAutoScalingGroupsResult

The AutoScalingGroupsType data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeAutoScalingGroupsResult.html

Constructors

DescribeAutoScalingGroupsResult

Fields

dasgrAutoScalingGroups :: Members AutoScalingGroup

A list of Auto Scaling groups.

dasgrNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

data DescribeAutoScalingInstancesResult

The AutoScalingInstancesType data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeAutoScalingInstancesResult.html

Constructors

DescribeAutoScalingInstancesResult

Fields

dasirAutoScalingInstances :: Members AutoScalingInstanceDetails

A list of Auto Scaling instances.

dasirNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

data DescribeAutoScalingNotificationTypesResult

The AutoScalingNotificationTypes data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeAutoScalingNotificationTypesResult.html

Constructors

DescribeAutoScalingNotificationTypesResult

Fields

dasntrAutoScalingNotificationTypes :: [Text]

Returns a list of all notification types supported by Auto Scaling.

data DescribeLaunchConfigurationsResult

The LaunchConfigurationsType data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeLaunchConfigurationsResult.html

Constructors

DescribeLaunchConfigurationsResult

Fields

dlcrLaunchConfigurations :: Members LaunchConfiguration

A list of launch configurations.

dlcrNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

data DescribeMetricCollectionTypesResult

The output of the DescribeMetricCollectionTypes action.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeMetricCollectionTypesResult.html

Constructors

DescribeMetricCollectionTypesResult

Fields

dmctrGranularities :: Members MetricGranularityType

A list of granularities for the listed Metrics.

dmctrMetrics :: Members MetricCollectionType

The list of Metrics collected.

data DescribeNotificationConfigurationsResult

The output of the DescribeNotificationConfigurations action.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeNotificationConfigurationsResult.html

Constructors

DescribeNotificationConfigurationsResult

Fields

dncrNextToken :: Maybe Text

A string that is used to mark the start of the next batch of returned results for pagination.

dncrNotificationConfigurations :: Members NotificationConfiguration

The list of notification configurations.

data DescribePoliciesResult

The PoliciesType data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribePoliciesResult.html

Constructors

DescribePoliciesResult

Fields

dprNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

dprScalingPolicies :: Members ScalingPolicy

A list of scaling policies.

data DescribeScalingActivitiesResult

The output for the DescribeScalingActivities action.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeScalingActivitiesResult.html

Constructors

DescribeScalingActivitiesResult

Fields

dsarActivities :: Members Activity

A list of the requested scaling activities.

dsarNextToken :: Maybe Text

Acts as a paging mechanism for large result sets. Set to a non-empty string if there are additional results waiting to be returned. Pass this in to subsequent calls to return additional results.

data DescribeScalingProcessTypesResult

The output of the DescribeScalingProcessTypes action.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeScalingProcessTypesResult.html

Constructors

DescribeScalingProcessTypesResult

Fields

dsptrProcesses :: Members ProcessType

A list of ProcessType names.

data DescribeScheduledActionsResult

A scaling action that is scheduled for a future time and date. An action can be scheduled up to thirty days in advance. Starting with API version 2011-01-01, you can use recurrence to specify that a scaling action occurs regularly on a schedule.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeScheduledActionsResult.html

Constructors

DescribeScheduledActionsResult

Fields

dsasNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

dsasScheduledUpdateGroupActions :: Members ScheduledUpdateGroupAction

A list of scheduled actions designed to update an Auto Scaling group.

data DescribeTagsResult

DescribeTagsResult

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeTagsResult.html

Constructors

DescribeTagsResult

Fields

dtrNextToken :: Maybe Text

A string used to mark the start of the next batch of returned results.

dtrTags :: Members Tag

The list of tags.

data DescribeTerminationPolicyTypesResult

The TerminationPolicyTypes data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeTerminationPolicyTypesResult.html

Constructors

DescribeTerminationPolicyTypesResult

Fields

dtptrTerminationPolicyTypes :: [Text]

Termination policies supported by Auto Scaling. They are: OldestInstance, OldestLaunchConfiguration, NewestInstance, ClosestToNextInstanceHour, Default

data Ebs

The Ebs data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_Ebs.html

Constructors

Ebs

Fields

eSnapshotId :: Maybe Text

The snapshot ID.

eVolumeSize :: Maybe Integer

The volume size, in gigabytes.

data EnabledMetric

The EnabledMetric data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_EnabledMetric.html

Constructors

EnabledMetric

Fields

emGranularity :: Maybe Text

The granularity of the enabled metric.

emMetric :: Maybe Text

The name of the enabled metric.

data Filter

The Filter data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_Filter.html

Constructors

Filter

Fields

fName :: Maybe Text

The name of the filter. Valid Name values are: auto-scaling-group, key, value, and propagate-at-launch.

fValues :: Maybe Text

The value of the filter.

data Instance

The Instance data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_Instance.html

Constructors

Instance

Fields

iAvailabilityZone :: !AvailabilityZone

Availability Zones associated with this instance.

iHealthStatus :: !Text

The instance's health status.

iInstanceId :: !Text

Specifies the ID of the Amazon EC2 instance.

iLaunchConfigurationName :: Maybe Text

The launch configuration associated with this instance.

iLifecycleState :: !Text

Contains a description of the current lifecycle state.

data InstanceMonitoring

The InstanceMonitoring data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_InstanceMonitoring.html

Constructors

InstanceMonitoring

Fields

imEnabled :: Bool

If True, instance monitoring is enabled.

data LaunchConfiguration

The LaunchConfiguration data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_LaunchConfiguration.html

Constructors

LaunchConfiguration

Fields

lcBlockDeviceMappings :: Members BlockDeviceMapping

Specifies how block devices are exposed to the instance. Each mapping is made up of a virtualName and a deviceName.

lcCreatedTime :: !UTCTime

Provides the creation date and time for this launch configuration.

lcEbsOptimized :: Maybe Bool

Specifies whether the instance is optimized for EBS I/O (true) or not (false).

lcIamInstanceProfile :: Maybe Text

Provides the name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

lcImageId :: !Text

Provides the unique ID of the Amazon Machine Image (AMI) that was assigned during registration.

lcInstanceMonitoring :: Maybe InstanceMonitoring

Controls whether instances in this group are launched with detailed monitoring or not.

lcInstanceType :: !Text

Specifies the instance type of the Amazon EC2 instance.

lcKernelId :: Maybe Text

Provides the ID of the kernel associated with the Amazon EC2 AMI.

lcKeyName :: Maybe Text

Provides the name of the Amazon EC2 key pair.

lcLaunchConfigurationARN :: Maybe Text

The launch configuration's Amazon Resource Name (ARN).

lcLaunchConfigurationName :: !Text

Specifies the name of the launch configuration.

lcRamdiskId :: Maybe Text

Provides ID of the RAM disk associated with the Amazon EC2 AMI.

lcSecurityGroups :: Maybe Text

A description of the security groups to associate with the Amazon EC2 instances.

lcSpotPrice :: Maybe Text

Specifies the price to bid when launching Spot Instances.

lcUserData :: Maybe Text

The user data available to the launched Amazon EC2 instances.

data MetricCollectionType

The MetricCollectionType data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_MetricCollectionType.html

Constructors

MetricCollectionType

Fields

mctMetric :: Text

Type: String

data MetricGranularityType

The MetricGranularityType data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_MetricGranularityType.html

Constructors

MetricGranularityType

Fields

mgtGranularity :: Text

The granularity of a Metric.

data NotificationConfiguration

The NotificationConfiguration data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_NotificationConfiguration.html

Constructors

NotificationConfiguration

Fields

ncAutoScalingGroupName :: Maybe Text

Specifies the Auto Scaling group name.

ncNotificationType :: Maybe Text

The types of events for an action to start.

ncTopicARN :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

data ProcessType

There are two primary Auto Scaling process types: Launch and Terminate. The Launch process creates a new Amazon EC2 instance for an Auto Scaling group, and the Terminate process removes an existing Amazon EC2 instance. The remaining Auto Scaling process types relate to specific Auto Scaling features: Important If you suspend Launch or Terminate, all other process types are affected to varying degrees.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_ProcessType.html

Constructors

ProcessType

Fields

ptProcessName :: !Text

The name of a process.

data PutScalingPolicyResult

The PolicyARNType data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_PutScalingPolicyResult.html

Constructors

PutScalingPolicyResult

Fields

psprPolicyARN :: Text

A policy's Amazon Resource Name (ARN).

data ScalingPolicy

The ScalingPolicy data type.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_ScalingPolicy.html

Constructors

ScalingPolicy

Fields

spAdjustmentType :: Maybe Text

Specifies whether the ScalingAdjustment is an absolute number or a percentage of the current capacity. Valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

spAlarms :: Members Alarm

A list of CloudWatch Alarms related to the policy.

sqAutoScalingGroupName :: Maybe Text

The name of the Auto Scaling group associated with this scaling policy.

sqCooldown :: Maybe Integer

The amount of time, in seconds, after a scaling activity completes before any further trigger-related scaling activities can start.

sqMinAdjustmentStep :: Maybe Integer

Changes the DesiredCapacity of the Auto Scaling group by at least the specified number of instances.

sqPolicyARN :: Maybe Text

The Amazon Resource Name (ARN) of the policy.

sqPolicyName :: Maybe Text

The name of the scaling policy.

sqScalingAdjustment :: Maybe Integer

The number associated with the specified adjustment type. A positive value adds to the current capacity and a negative value removes from the current capacity.

data ScheduledUpdateGroupAction

This data type stores information about a scheduled update to an Auto Scaling group.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_ScheduledUpdateGroupAction.html

Constructors

ScheduledUpdateGroupAction

Fields

sugaAutoScalingGroupName :: Maybe Text

The name of the Auto Scaling group to be updated.

sugaDesiredCapacity :: Maybe Integer

The number of instances you prefer to maintain in your Auto Scaling group.

sugaEndTime :: Maybe UTCTime

The time that the action is scheduled to end. This value can be up to one month in the future.

sugaMaxSize :: Maybe Integer

The maximum size of the Auto Scaling group.

sugaMinSize :: Maybe Integer

The minimum size of the Auto Scaling group.

sugaRecurrence :: Maybe Text

The regular schedule that an action occurs.

sugaScheduledActionARN :: Maybe Text

The Amazon Resource Name (ARN) of this scheduled action.

sugaScheduledActionName :: Maybe Text

The name of this scheduled action.

sugaStartTime :: Maybe UTCTime

The time that the action is scheduled to begin. This value can be up to one month in the future.

sugaTime :: Maybe UTCTime

Time is deprecated.

data SuspendedProcess

An Auto Scaling process that has been suspended. For more information, see ProcessType.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_SuspendedProcess.html

Constructors

SuspendedProcess

Fields

spProcessName :: Maybe Text

The name of the suspended process.

spSuspensionReason :: Maybe Text

The reason that the process was suspended.

data Tag

The tag applied to an Auto Scaling group.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_Tag.html

Constructors

Tag

Fields

tKey :: !Text

The key of the tag.

tPropagateAtLaunch :: Maybe Bool

Specifies whether the new tag will be applied to instances launched after the tag is created. The same behavior applies to updates: If you change a tag, the changed tag will be applied to all instances launched after you made the change.

tResourceId :: Maybe Text

The name of the Auto Scaling group.

tResourceType :: Maybe Text

The kind of resource to which the tag is applied. Currently, Auto Scaling supports the auto-scaling-group resource type.

tValue :: Maybe Text

The value of the tag.

data TerminateInstanceInAutoScalingGroupResult

The output for the TerminateInstanceInAutoScalingGroup action.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_TerminateInstanceInAutoScalingGroupResult.html

Constructors

TerminateInstanceInAutoScalingGroupResult

Fields

tiiasgrActivity :: Maybe Activity

A scaling Activity.

autoScalingNS :: ByteString

XML namespace to annotate AutoScaling elements with.

autoScalingElem :: ByteString -> NName ByteString

Helper to define AutoScaling namespaced XML elements.

Common

module Network.AWS