Network.AWS.AutoScaling.AttachInstances

Description

Attaches one or more EC2 instances to the specified Auto Scaling group.

When you attach instances, Amazon EC2 Auto Scaling increases the desired capacity of the group by the number of instances being attached. If the number of instances being attached plus the desired capacity of the group exceeds the maximum size of the group, the operation fails.

If there is a Classic Load Balancer attached to your Auto Scaling group, the instances are also registered with the load balancer. If there are target groups attached to your Auto Scaling group, the instances are also registered with the target groups.

For more information, see Attach EC2 instances to your Auto Scaling group in the Amazon EC2 Auto Scaling User Guide.

Creating a Request

data AttachInstances

See: newAttachInstances smart constructor.

Constructors

AttachInstances'

Fields

instanceIds :: Maybe [Text]

The IDs of the instances. You can specify up to 20 instances.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

newAttachInstances

Arguments

:: Text
$sel:autoScalingGroupName:AttachInstances'

-> AttachInstances

Create a value of AttachInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceIds:AttachInstances', attachInstances_instanceIds - The IDs of the instances. You can specify up to 20 instances.

$sel:autoScalingGroupName:AttachInstances', attachInstances_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

attachInstances_instanceIds :: Lens' AttachInstances (Maybe [Text])

The IDs of the instances. You can specify up to 20 instances.

attachInstances_autoScalingGroupName :: Lens' AttachInstances Text

The name of the Auto Scaling group.

Destructuring the Response

data AttachInstancesResponse

See: newAttachInstancesResponse smart constructor.

Constructors

AttachInstancesResponse'

newAttachInstancesResponse :: AttachInstancesResponse

Create a value of AttachInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.