Network.AWS.AutoScaling.CreateLaunchConfiguration

Description

Creates a launch configuration.

If you exceed your maximum limit of launch configurations, the call fails. To query this limit, call the DescribeAccountLimits API. For information about updating this limit, see Amazon EC2 Auto Scaling service quotas in the Amazon EC2 Auto Scaling User Guide.

For more information, see Launch configurations in the Amazon EC2 Auto Scaling User Guide.

Creating a Request

data CreateLaunchConfiguration

See: newCreateLaunchConfiguration smart constructor.

Constructors

CreateLaunchConfiguration'

Fields

instanceId :: Maybe Text

The ID of the instance to use to create the launch configuration. The new launch configuration derives attributes from the instance, except for the block device mapping.

To create a launch configuration with a block device mapping or override any other instance attributes, specify them as part of the same request.

For more information, see Creating a launch configuration using an EC2 instance in the Amazon EC2 Auto Scaling User Guide.

If you do not specify InstanceId, you must specify both ImageId and InstanceType.

ebsOptimized :: Maybe Bool

Specifies whether the launch configuration is optimized for EBS I/O (true) or not (false). The optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization is not available with all instance types. Additional fees are incurred when you enable EBS optimization for an instance type that is not EBS-optimized by default. For more information, see Amazon EBS-Optimized Instances in the Amazon EC2 User Guide for Linux Instances.

The default value is false.

userData :: Maybe Text

The user data to make available to the launched EC2 instances. For more information, see Instance metadata and user data (Linux) and Instance metadata and user data (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text. User data is limited to 16 KB.

instanceType :: Maybe Text

Specifies the instance type of the EC2 instance.

For information about available instance types, see Available Instance Types in the Amazon EC2 User Guide for Linux Instances.

If you do not specify InstanceId, you must specify InstanceType.

ramdiskId :: Maybe Text

The ID of the RAM disk to select.

classicLinkVPCSecurityGroups :: Maybe [Text]

The IDs of one or more security groups for the specified ClassicLink-enabled VPC. For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

If you specify the ClassicLinkVPCId parameter, you must specify this parameter.

spotPrice :: Maybe Text

The maximum hourly price to be paid for any Spot Instance launched to fulfill the request. Spot Instances are launched when the price you specify exceeds the current Spot price. For more information, see Requesting Spot Instances in the Amazon EC2 Auto Scaling User Guide.

When you change your maximum price by creating a new launch configuration, running instances will continue to run as long as the maximum price for those running instances is higher than the current Spot price.

associatePublicIpAddress :: Maybe Bool

For Auto Scaling groups that are running in a virtual private cloud (VPC), specifies whether to assign a public IP address to the group's instances. If you specify true, each instance in the Auto Scaling group receives a unique public IP address. For more information, see Launching Auto Scaling instances in a VPC in the Amazon EC2 Auto Scaling User Guide.

If you specify this parameter, you must specify at least one subnet for VPCZoneIdentifier when you create your group.

If the instance is launched into a default subnet, the default is to assign a public IP address, unless you disabled the option to assign a public IP address on the subnet. If the instance is launched into a nondefault subnet, the default is not to assign a public IP address, unless you enabled the option to assign a public IP address on the subnet.

imageId :: Maybe Text

The ID of the Amazon Machine Image (AMI) that was assigned during registration. For more information, see Finding an AMI in the Amazon EC2 User Guide for Linux Instances.

If you do not specify InstanceId, you must specify ImageId.

securityGroups :: Maybe [Text]

A list that contains the security groups to assign to the instances in the Auto Scaling group.

EC2-VPC

Specify the security group IDs. For more information, see Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide.

EC2-Classic

Specify either the security group names or the security group IDs. For more information, see Amazon EC2 Security Groups in the Amazon EC2 User Guide for Linux Instances.

iamInstanceProfile :: Maybe Text

The name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

For more information, see IAM role for applications that run on Amazon EC2 instances in the Amazon EC2 Auto Scaling User Guide.

classicLinkVPCId :: Maybe Text

The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

This parameter can only be used if you are launching EC2-Classic instances.

blockDeviceMappings :: Maybe [BlockDeviceMapping]

A block device mapping, which specifies the block devices for the instance. You can specify virtual devices and EBS volumes. For more information, see Block Device Mapping in the Amazon EC2 User Guide for Linux Instances.

kernelId :: Maybe Text

The ID of the kernel associated with the AMI.

placementTenancy :: Maybe Text

The tenancy of the instance. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

To launch dedicated instances into a shared tenancy VPC (a VPC with the instance placement tenancy attribute set to default), you must set the value of this parameter to dedicated.

If you specify PlacementTenancy, you must specify at least one subnet for VPCZoneIdentifier when you create your group.

For more information, see Configuring instance tenancy with Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

Valid Values: default | dedicated

keyName :: Maybe Text

The name of the key pair. For more information, see Amazon EC2 Key Pairs in the Amazon EC2 User Guide for Linux Instances.

instanceMonitoring :: Maybe InstanceMonitoring

Controls whether instances in this group are launched with detailed (true) or basic (false) monitoring.

The default value is true (enabled).

When detailed monitoring is enabled, Amazon CloudWatch generates metrics every minute and your account is charged a fee. When you disable detailed monitoring, CloudWatch generates metrics every 5 minutes. For more information, see Configure Monitoring for Auto Scaling Instances in the Amazon EC2 Auto Scaling User Guide.

metadataOptions :: Maybe InstanceMetadataOptions

The metadata options for the instances. For more information, see Configuring the Instance Metadata Options in the Amazon EC2 Auto Scaling User Guide.

launchConfigurationName :: Text

The name of the launch configuration. This name must be unique per Region per account.

newCreateLaunchConfiguration

Arguments

:: Text
$sel:launchConfigurationName:CreateLaunchConfiguration'

-> CreateLaunchConfiguration

Create a value of CreateLaunchConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:CreateLaunchConfiguration', createLaunchConfiguration_instanceId - The ID of the instance to use to create the launch configuration. The new launch configuration derives attributes from the instance, except for the block device mapping.

To create a launch configuration with a block device mapping or override any other instance attributes, specify them as part of the same request.

For more information, see Creating a launch configuration using an EC2 instance in the Amazon EC2 Auto Scaling User Guide.

If you do not specify InstanceId, you must specify both ImageId and InstanceType.

$sel:ebsOptimized:CreateLaunchConfiguration', createLaunchConfiguration_ebsOptimized - Specifies whether the launch configuration is optimized for EBS I/O (true) or not (false). The optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization is not available with all instance types. Additional fees are incurred when you enable EBS optimization for an instance type that is not EBS-optimized by default. For more information, see Amazon EBS-Optimized Instances in the Amazon EC2 User Guide for Linux Instances.

The default value is false.

$sel:userData:CreateLaunchConfiguration', createLaunchConfiguration_userData - The user data to make available to the launched EC2 instances. For more information, see Instance metadata and user data (Linux) and Instance metadata and user data (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text. User data is limited to 16 KB.

$sel:instanceType:CreateLaunchConfiguration', createLaunchConfiguration_instanceType - Specifies the instance type of the EC2 instance.

For information about available instance types, see Available Instance Types in the Amazon EC2 User Guide for Linux Instances.

If you do not specify InstanceId, you must specify InstanceType.

$sel:ramdiskId:CreateLaunchConfiguration', createLaunchConfiguration_ramdiskId - The ID of the RAM disk to select.

$sel:classicLinkVPCSecurityGroups:CreateLaunchConfiguration', createLaunchConfiguration_classicLinkVPCSecurityGroups - The IDs of one or more security groups for the specified ClassicLink-enabled VPC. For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

If you specify the ClassicLinkVPCId parameter, you must specify this parameter.

$sel:spotPrice:CreateLaunchConfiguration', createLaunchConfiguration_spotPrice - The maximum hourly price to be paid for any Spot Instance launched to fulfill the request. Spot Instances are launched when the price you specify exceeds the current Spot price. For more information, see Requesting Spot Instances in the Amazon EC2 Auto Scaling User Guide.

When you change your maximum price by creating a new launch configuration, running instances will continue to run as long as the maximum price for those running instances is higher than the current Spot price.

$sel:associatePublicIpAddress:CreateLaunchConfiguration', createLaunchConfiguration_associatePublicIpAddress - For Auto Scaling groups that are running in a virtual private cloud (VPC), specifies whether to assign a public IP address to the group's instances. If you specify true, each instance in the Auto Scaling group receives a unique public IP address. For more information, see Launching Auto Scaling instances in a VPC in the Amazon EC2 Auto Scaling User Guide.

If you specify this parameter, you must specify at least one subnet for VPCZoneIdentifier when you create your group.

If the instance is launched into a default subnet, the default is to assign a public IP address, unless you disabled the option to assign a public IP address on the subnet. If the instance is launched into a nondefault subnet, the default is not to assign a public IP address, unless you enabled the option to assign a public IP address on the subnet.

$sel:imageId:CreateLaunchConfiguration', createLaunchConfiguration_imageId - The ID of the Amazon Machine Image (AMI) that was assigned during registration. For more information, see Finding an AMI in the Amazon EC2 User Guide for Linux Instances.

If you do not specify InstanceId, you must specify ImageId.

$sel:securityGroups:CreateLaunchConfiguration', createLaunchConfiguration_securityGroups - A list that contains the security groups to assign to the instances in the Auto Scaling group.

EC2-VPC

Specify the security group IDs. For more information, see Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide.

EC2-Classic

Specify either the security group names or the security group IDs. For more information, see Amazon EC2 Security Groups in the Amazon EC2 User Guide for Linux Instances.

$sel:iamInstanceProfile:CreateLaunchConfiguration', createLaunchConfiguration_iamInstanceProfile - The name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

For more information, see IAM role for applications that run on Amazon EC2 instances in the Amazon EC2 Auto Scaling User Guide.

$sel:classicLinkVPCId:CreateLaunchConfiguration', createLaunchConfiguration_classicLinkVPCId - The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

This parameter can only be used if you are launching EC2-Classic instances.

$sel:blockDeviceMappings:CreateLaunchConfiguration', createLaunchConfiguration_blockDeviceMappings - A block device mapping, which specifies the block devices for the instance. You can specify virtual devices and EBS volumes. For more information, see Block Device Mapping in the Amazon EC2 User Guide for Linux Instances.

$sel:kernelId:CreateLaunchConfiguration', createLaunchConfiguration_kernelId - The ID of the kernel associated with the AMI.

$sel:placementTenancy:CreateLaunchConfiguration', createLaunchConfiguration_placementTenancy - The tenancy of the instance. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

To launch dedicated instances into a shared tenancy VPC (a VPC with the instance placement tenancy attribute set to default), you must set the value of this parameter to dedicated.

If you specify PlacementTenancy, you must specify at least one subnet for VPCZoneIdentifier when you create your group.

For more information, see Configuring instance tenancy with Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

Valid Values: default | dedicated

$sel:keyName:CreateLaunchConfiguration', createLaunchConfiguration_keyName - The name of the key pair. For more information, see Amazon EC2 Key Pairs in the Amazon EC2 User Guide for Linux Instances.

$sel:instanceMonitoring:CreateLaunchConfiguration', createLaunchConfiguration_instanceMonitoring - Controls whether instances in this group are launched with detailed (true) or basic (false) monitoring.

The default value is true (enabled).

When detailed monitoring is enabled, Amazon CloudWatch generates metrics every minute and your account is charged a fee. When you disable detailed monitoring, CloudWatch generates metrics every 5 minutes. For more information, see Configure Monitoring for Auto Scaling Instances in the Amazon EC2 Auto Scaling User Guide.

$sel:metadataOptions:CreateLaunchConfiguration', createLaunchConfiguration_metadataOptions - The metadata options for the instances. For more information, see Configuring the Instance Metadata Options in the Amazon EC2 Auto Scaling User Guide.

$sel:launchConfigurationName:CreateLaunchConfiguration', createLaunchConfiguration_launchConfigurationName - The name of the launch configuration. This name must be unique per Region per account.

Request Lenses

createLaunchConfiguration_instanceId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of the instance to use to create the launch configuration. The new launch configuration derives attributes from the instance, except for the block device mapping.

To create a launch configuration with a block device mapping or override any other instance attributes, specify them as part of the same request.

For more information, see Creating a launch configuration using an EC2 instance in the Amazon EC2 Auto Scaling User Guide.

If you do not specify InstanceId, you must specify both ImageId and InstanceType.

createLaunchConfiguration_ebsOptimized :: Lens' CreateLaunchConfiguration (Maybe Bool)

Specifies whether the launch configuration is optimized for EBS I/O (true) or not (false). The optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization is not available with all instance types. Additional fees are incurred when you enable EBS optimization for an instance type that is not EBS-optimized by default. For more information, see Amazon EBS-Optimized Instances in the Amazon EC2 User Guide for Linux Instances.

The default value is false.

createLaunchConfiguration_userData :: Lens' CreateLaunchConfiguration (Maybe Text)

The user data to make available to the launched EC2 instances. For more information, see Instance metadata and user data (Linux) and Instance metadata and user data (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text. User data is limited to 16 KB.

createLaunchConfiguration_instanceType :: Lens' CreateLaunchConfiguration (Maybe Text)

Specifies the instance type of the EC2 instance.

For information about available instance types, see Available Instance Types in the Amazon EC2 User Guide for Linux Instances.

If you do not specify InstanceId, you must specify InstanceType.

createLaunchConfiguration_ramdiskId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of the RAM disk to select.

createLaunchConfiguration_classicLinkVPCSecurityGroups :: Lens' CreateLaunchConfiguration (Maybe [Text])

The IDs of one or more security groups for the specified ClassicLink-enabled VPC. For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

If you specify the ClassicLinkVPCId parameter, you must specify this parameter.

createLaunchConfiguration_spotPrice :: Lens' CreateLaunchConfiguration (Maybe Text)

The maximum hourly price to be paid for any Spot Instance launched to fulfill the request. Spot Instances are launched when the price you specify exceeds the current Spot price. For more information, see Requesting Spot Instances in the Amazon EC2 Auto Scaling User Guide.

When you change your maximum price by creating a new launch configuration, running instances will continue to run as long as the maximum price for those running instances is higher than the current Spot price.

createLaunchConfiguration_associatePublicIpAddress :: Lens' CreateLaunchConfiguration (Maybe Bool)

For Auto Scaling groups that are running in a virtual private cloud (VPC), specifies whether to assign a public IP address to the group's instances. If you specify true, each instance in the Auto Scaling group receives a unique public IP address. For more information, see Launching Auto Scaling instances in a VPC in the Amazon EC2 Auto Scaling User Guide.

If you specify this parameter, you must specify at least one subnet for VPCZoneIdentifier when you create your group.

If the instance is launched into a default subnet, the default is to assign a public IP address, unless you disabled the option to assign a public IP address on the subnet. If the instance is launched into a nondefault subnet, the default is not to assign a public IP address, unless you enabled the option to assign a public IP address on the subnet.

createLaunchConfiguration_imageId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of the Amazon Machine Image (AMI) that was assigned during registration. For more information, see Finding an AMI in the Amazon EC2 User Guide for Linux Instances.

If you do not specify InstanceId, you must specify ImageId.

createLaunchConfiguration_securityGroups :: Lens' CreateLaunchConfiguration (Maybe [Text])

A list that contains the security groups to assign to the instances in the Auto Scaling group.

EC2-VPC

Specify the security group IDs. For more information, see Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide.

EC2-Classic

Specify either the security group names or the security group IDs. For more information, see Amazon EC2 Security Groups in the Amazon EC2 User Guide for Linux Instances.

createLaunchConfiguration_iamInstanceProfile :: Lens' CreateLaunchConfiguration (Maybe Text)

The name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

For more information, see IAM role for applications that run on Amazon EC2 instances in the Amazon EC2 Auto Scaling User Guide.

createLaunchConfiguration_classicLinkVPCId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

This parameter can only be used if you are launching EC2-Classic instances.

createLaunchConfiguration_blockDeviceMappings :: Lens' CreateLaunchConfiguration (Maybe [BlockDeviceMapping])

A block device mapping, which specifies the block devices for the instance. You can specify virtual devices and EBS volumes. For more information, see Block Device Mapping in the Amazon EC2 User Guide for Linux Instances.

createLaunchConfiguration_kernelId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of the kernel associated with the AMI.

createLaunchConfiguration_placementTenancy :: Lens' CreateLaunchConfiguration (Maybe Text)

The tenancy of the instance. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

To launch dedicated instances into a shared tenancy VPC (a VPC with the instance placement tenancy attribute set to default), you must set the value of this parameter to dedicated.

If you specify PlacementTenancy, you must specify at least one subnet for VPCZoneIdentifier when you create your group.

For more information, see Configuring instance tenancy with Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

Valid Values: default | dedicated

createLaunchConfiguration_keyName :: Lens' CreateLaunchConfiguration (Maybe Text)

The name of the key pair. For more information, see Amazon EC2 Key Pairs in the Amazon EC2 User Guide for Linux Instances.

createLaunchConfiguration_instanceMonitoring :: Lens' CreateLaunchConfiguration (Maybe InstanceMonitoring)

Controls whether instances in this group are launched with detailed (true) or basic (false) monitoring.

The default value is true (enabled).

When detailed monitoring is enabled, Amazon CloudWatch generates metrics every minute and your account is charged a fee. When you disable detailed monitoring, CloudWatch generates metrics every 5 minutes. For more information, see Configure Monitoring for Auto Scaling Instances in the Amazon EC2 Auto Scaling User Guide.

createLaunchConfiguration_metadataOptions :: Lens' CreateLaunchConfiguration (Maybe InstanceMetadataOptions)

The metadata options for the instances. For more information, see Configuring the Instance Metadata Options in the Amazon EC2 Auto Scaling User Guide.

createLaunchConfiguration_launchConfigurationName :: Lens' CreateLaunchConfiguration Text

The name of the launch configuration. This name must be unique per Region per account.

Destructuring the Response

data CreateLaunchConfigurationResponse

See: newCreateLaunchConfigurationResponse smart constructor.

Constructors

CreateLaunchConfigurationResponse'

newCreateLaunchConfigurationResponse :: CreateLaunchConfigurationResponse

Create a value of CreateLaunchConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.