Network.AWS.AutoScaling.DeleteAutoScalingGroup

Description

Deletes the specified Auto Scaling group.

If the group has instances or scaling activities in progress, you must specify the option to force the deletion in order for it to succeed.

If the group has policies, deleting the group deletes the policies, the underlying alarm actions, and any alarm that no longer has an associated action.

To remove instances from the Auto Scaling group before deleting it, call the DetachInstances API with the list of instances and the option to decrement the desired capacity. This ensures that Amazon EC2 Auto Scaling does not launch replacement instances.

To terminate all instances before deleting the Auto Scaling group, call the UpdateAutoScalingGroup API and set the minimum size and desired capacity of the Auto Scaling group to zero.

Creating a Request

data DeleteAutoScalingGroup

See: newDeleteAutoScalingGroup smart constructor.

Constructors

DeleteAutoScalingGroup'

Fields

forceDelete :: Maybe Bool

Specifies that the group is to be deleted along with all instances associated with the group, without waiting for all instances to be terminated. This parameter also deletes any outstanding lifecycle actions associated with the group.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

newDeleteAutoScalingGroup

Arguments

:: Text
$sel:autoScalingGroupName:DeleteAutoScalingGroup'

-> DeleteAutoScalingGroup

Create a value of DeleteAutoScalingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forceDelete:DeleteAutoScalingGroup', deleteAutoScalingGroup_forceDelete - Specifies that the group is to be deleted along with all instances associated with the group, without waiting for all instances to be terminated. This parameter also deletes any outstanding lifecycle actions associated with the group.

$sel:autoScalingGroupName:DeleteAutoScalingGroup', deleteAutoScalingGroup_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

deleteAutoScalingGroup_forceDelete :: Lens' DeleteAutoScalingGroup (Maybe Bool)

Specifies that the group is to be deleted along with all instances associated with the group, without waiting for all instances to be terminated. This parameter also deletes any outstanding lifecycle actions associated with the group.

deleteAutoScalingGroup_autoScalingGroupName :: Lens' DeleteAutoScalingGroup Text

The name of the Auto Scaling group.

Destructuring the Response

data DeleteAutoScalingGroupResponse

See: newDeleteAutoScalingGroupResponse smart constructor.

Constructors

DeleteAutoScalingGroupResponse'

newDeleteAutoScalingGroupResponse :: DeleteAutoScalingGroupResponse

Create a value of DeleteAutoScalingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.