Network.AWS.AutoScaling.DescribeAccountLimits

Description

Describes the current Amazon EC2 Auto Scaling resource quotas for your account.

When you establish an account, the account has initial quotas on the maximum number of Auto Scaling groups and launch configurations that you can create in a given Region. For more information, see Amazon EC2 Auto Scaling service quotas in the Amazon EC2 Auto Scaling User Guide.

Creating a Request

data DescribeAccountLimits

See: newDescribeAccountLimits smart constructor.

Constructors

DescribeAccountLimits'

newDescribeAccountLimits :: DescribeAccountLimits

Create a value of DescribeAccountLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeAccountLimitsResponse

See: newDescribeAccountLimitsResponse smart constructor.

Constructors

DescribeAccountLimitsResponse'

Fields

numberOfAutoScalingGroups :: Maybe Int

The current number of groups for your account.

maxNumberOfLaunchConfigurations :: Maybe Int

The maximum number of launch configurations allowed for your account. The default is 200 launch configurations per Region.

numberOfLaunchConfigurations :: Maybe Int

The current number of launch configurations for your account.

maxNumberOfAutoScalingGroups :: Maybe Int

The maximum number of groups allowed for your account. The default is 200 groups per Region.

httpStatus :: Int

The response's http status code.

newDescribeAccountLimitsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAccountLimitsResponse'

-> DescribeAccountLimitsResponse

Create a value of DescribeAccountLimitsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numberOfAutoScalingGroups:DescribeAccountLimitsResponse', describeAccountLimitsResponse_numberOfAutoScalingGroups - The current number of groups for your account.

$sel:maxNumberOfLaunchConfigurations:DescribeAccountLimitsResponse', describeAccountLimitsResponse_maxNumberOfLaunchConfigurations - The maximum number of launch configurations allowed for your account. The default is 200 launch configurations per Region.

$sel:numberOfLaunchConfigurations:DescribeAccountLimitsResponse', describeAccountLimitsResponse_numberOfLaunchConfigurations - The current number of launch configurations for your account.

$sel:maxNumberOfAutoScalingGroups:DescribeAccountLimitsResponse', describeAccountLimitsResponse_maxNumberOfAutoScalingGroups - The maximum number of groups allowed for your account. The default is 200 groups per Region.

$sel:httpStatus:DescribeAccountLimitsResponse', describeAccountLimitsResponse_httpStatus - The response's http status code.

Response Lenses

describeAccountLimitsResponse_numberOfAutoScalingGroups :: Lens' DescribeAccountLimitsResponse (Maybe Int)

The current number of groups for your account.

describeAccountLimitsResponse_maxNumberOfLaunchConfigurations :: Lens' DescribeAccountLimitsResponse (Maybe Int)

The maximum number of launch configurations allowed for your account. The default is 200 launch configurations per Region.

describeAccountLimitsResponse_numberOfLaunchConfigurations :: Lens' DescribeAccountLimitsResponse (Maybe Int)

The current number of launch configurations for your account.

describeAccountLimitsResponse_maxNumberOfAutoScalingGroups :: Lens' DescribeAccountLimitsResponse (Maybe Int)

The maximum number of groups allowed for your account. The default is 200 groups per Region.

describeAccountLimitsResponse_httpStatus :: Lens' DescribeAccountLimitsResponse Int

The response's http status code.