Network.AWS.AutoScaling.DescribeLifecycleHooks

Description

Gets information about the lifecycle hooks for the specified Auto Scaling group.

Creating a Request

data DescribeLifecycleHooks

See: newDescribeLifecycleHooks smart constructor.

Constructors

DescribeLifecycleHooks'

Fields

lifecycleHookNames :: Maybe [Text]

The names of one or more lifecycle hooks. If you omit this parameter, all lifecycle hooks are described.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

newDescribeLifecycleHooks

Arguments

:: Text
$sel:autoScalingGroupName:DescribeLifecycleHooks'

-> DescribeLifecycleHooks

Create a value of DescribeLifecycleHooks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lifecycleHookNames:DescribeLifecycleHooks', describeLifecycleHooks_lifecycleHookNames - The names of one or more lifecycle hooks. If you omit this parameter, all lifecycle hooks are described.

$sel:autoScalingGroupName:DescribeLifecycleHooks', describeLifecycleHooks_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

describeLifecycleHooks_lifecycleHookNames :: Lens' DescribeLifecycleHooks (Maybe [Text])

The names of one or more lifecycle hooks. If you omit this parameter, all lifecycle hooks are described.

describeLifecycleHooks_autoScalingGroupName :: Lens' DescribeLifecycleHooks Text

The name of the Auto Scaling group.

Destructuring the Response

data DescribeLifecycleHooksResponse

See: newDescribeLifecycleHooksResponse smart constructor.

Constructors

DescribeLifecycleHooksResponse'

Fields

lifecycleHooks :: Maybe [LifecycleHook]

The lifecycle hooks for the specified group.

httpStatus :: Int

The response's http status code.

newDescribeLifecycleHooksResponse

Arguments

:: Int
$sel:httpStatus:DescribeLifecycleHooksResponse'

-> DescribeLifecycleHooksResponse

Create a value of DescribeLifecycleHooksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lifecycleHooks:DescribeLifecycleHooksResponse', describeLifecycleHooksResponse_lifecycleHooks - The lifecycle hooks for the specified group.

$sel:httpStatus:DescribeLifecycleHooksResponse', describeLifecycleHooksResponse_httpStatus - The response's http status code.

Response Lenses

describeLifecycleHooksResponse_lifecycleHooks :: Lens' DescribeLifecycleHooksResponse (Maybe [LifecycleHook])

The lifecycle hooks for the specified group.

describeLifecycleHooksResponse_httpStatus :: Lens' DescribeLifecycleHooksResponse Int

The response's http status code.