Network.AWS.AutoScaling.DescribeLoadBalancers

Description

Gets information about the load balancers for the specified Auto Scaling group.

This operation describes only Classic Load Balancers. If you have Application Load Balancers, Network Load Balancers, or Gateway Load Balancers, use the DescribeLoadBalancerTargetGroups API instead.

To determine the availability of registered instances, use the State element in the response. When you attach a load balancer to an Auto Scaling group, the initial State value is Adding. The state transitions to Added after all Auto Scaling instances are registered with the load balancer. If Elastic Load Balancing health checks are enabled for the Auto Scaling group, the state transitions to InService after at least one Auto Scaling instance passes the health check. When the load balancer is in the InService state, Amazon EC2 Auto Scaling can terminate and replace any instances that are reported as unhealthy. If no registered instances pass the health checks, the load balancer doesn't enter the InService state.

Load balancers also have an InService state if you attach them in the CreateAutoScalingGroup API call. If your load balancer state is InService, but it is not working properly, check the scaling activities by calling DescribeScalingActivities and take any corrective actions necessary.

For help with failed health checks, see Troubleshooting Amazon EC2 Auto Scaling: Health checks in the Amazon EC2 Auto Scaling User Guide. For more information, see Elastic Load Balancing and Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

This operation returns paginated results.

Creating a Request

data DescribeLoadBalancers

See: newDescribeLoadBalancers smart constructor.

Constructors

DescribeLoadBalancers'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

maxRecords :: Maybe Int

The maximum number of items to return with this call. The default value is 100 and the maximum value is 100.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

newDescribeLoadBalancers

Arguments

:: Text
$sel:autoScalingGroupName:DescribeLoadBalancers'

-> DescribeLoadBalancers

Create a value of DescribeLoadBalancers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeLoadBalancers', describeLoadBalancers_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:maxRecords:DescribeLoadBalancers', describeLoadBalancers_maxRecords - The maximum number of items to return with this call. The default value is 100 and the maximum value is 100.

$sel:autoScalingGroupName:DescribeLoadBalancers', describeLoadBalancers_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

describeLoadBalancers_nextToken :: Lens' DescribeLoadBalancers (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeLoadBalancers_maxRecords :: Lens' DescribeLoadBalancers (Maybe Int)

The maximum number of items to return with this call. The default value is 100 and the maximum value is 100.

describeLoadBalancers_autoScalingGroupName :: Lens' DescribeLoadBalancers Text

The name of the Auto Scaling group.

Destructuring the Response

data DescribeLoadBalancersResponse

See: newDescribeLoadBalancersResponse smart constructor.

Constructors

DescribeLoadBalancersResponse'

Fields

nextToken :: Maybe Text

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

loadBalancers :: Maybe [LoadBalancerState]

The load balancers.

httpStatus :: Int

The response's http status code.

newDescribeLoadBalancersResponse

Arguments

:: Int
$sel:httpStatus:DescribeLoadBalancersResponse'

-> DescribeLoadBalancersResponse

Create a value of DescribeLoadBalancersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeLoadBalancers', describeLoadBalancersResponse_nextToken - A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

$sel:loadBalancers:DescribeLoadBalancersResponse', describeLoadBalancersResponse_loadBalancers - The load balancers.

$sel:httpStatus:DescribeLoadBalancersResponse', describeLoadBalancersResponse_httpStatus - The response's http status code.

Response Lenses

describeLoadBalancersResponse_nextToken :: Lens' DescribeLoadBalancersResponse (Maybe Text)

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

describeLoadBalancersResponse_loadBalancers :: Lens' DescribeLoadBalancersResponse (Maybe [LoadBalancerState])

The load balancers.

describeLoadBalancersResponse_httpStatus :: Lens' DescribeLoadBalancersResponse Int

The response's http status code.