Network.AWS.AutoScaling.DescribePolicies

Description

Gets information about the scaling policies in the account and Region.

This operation returns paginated results.

Creating a Request

data DescribePolicies

See: newDescribePolicies smart constructor.

Constructors

DescribePolicies'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

policyTypes :: Maybe [Text]

One or more policy types. The valid values are SimpleScaling, StepScaling, TargetTrackingScaling, and PredictiveScaling.

policyNames :: Maybe [Text]

The names of one or more policies. If you omit this parameter, all policies are described. If a group name is provided, the results are limited to that group. If you specify an unknown policy name, it is ignored with no error.

Array Members: Maximum number of 50 items.

maxRecords :: Maybe Int

The maximum number of items to be returned with each call. The default value is 50 and the maximum value is 100.

autoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

newDescribePolicies :: DescribePolicies

Create a value of DescribePolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribePolicies', describePolicies_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:policyTypes:DescribePolicies', describePolicies_policyTypes - One or more policy types. The valid values are SimpleScaling, StepScaling, TargetTrackingScaling, and PredictiveScaling.

$sel:policyNames:DescribePolicies', describePolicies_policyNames - The names of one or more policies. If you omit this parameter, all policies are described. If a group name is provided, the results are limited to that group. If you specify an unknown policy name, it is ignored with no error.

Array Members: Maximum number of 50 items.

$sel:maxRecords:DescribePolicies', describePolicies_maxRecords - The maximum number of items to be returned with each call. The default value is 50 and the maximum value is 100.

$sel:autoScalingGroupName:DescribePolicies', describePolicies_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

describePolicies_nextToken :: Lens' DescribePolicies (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describePolicies_policyTypes :: Lens' DescribePolicies (Maybe [Text])

One or more policy types. The valid values are SimpleScaling, StepScaling, TargetTrackingScaling, and PredictiveScaling.

describePolicies_policyNames :: Lens' DescribePolicies (Maybe [Text])

The names of one or more policies. If you omit this parameter, all policies are described. If a group name is provided, the results are limited to that group. If you specify an unknown policy name, it is ignored with no error.

Array Members: Maximum number of 50 items.

describePolicies_maxRecords :: Lens' DescribePolicies (Maybe Int)

The maximum number of items to be returned with each call. The default value is 50 and the maximum value is 100.

describePolicies_autoScalingGroupName :: Lens' DescribePolicies (Maybe Text)

The name of the Auto Scaling group.

Destructuring the Response

data DescribePoliciesResponse

See: newDescribePoliciesResponse smart constructor.

Constructors

DescribePoliciesResponse'

Fields

nextToken :: Maybe Text

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

scalingPolicies :: Maybe [ScalingPolicy]

The scaling policies.

httpStatus :: Int

The response's http status code.

newDescribePoliciesResponse

Arguments

:: Int
$sel:httpStatus:DescribePoliciesResponse'

-> DescribePoliciesResponse

Create a value of DescribePoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribePolicies', describePoliciesResponse_nextToken - A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

$sel:scalingPolicies:DescribePoliciesResponse', describePoliciesResponse_scalingPolicies - The scaling policies.

$sel:httpStatus:DescribePoliciesResponse', describePoliciesResponse_httpStatus - The response's http status code.

Response Lenses

describePoliciesResponse_nextToken :: Lens' DescribePoliciesResponse (Maybe Text)

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

describePoliciesResponse_scalingPolicies :: Lens' DescribePoliciesResponse (Maybe [ScalingPolicy])

The scaling policies.

describePoliciesResponse_httpStatus :: Lens' DescribePoliciesResponse Int

The response's http status code.