Network.AWS.AutoScaling.DescribeScalingActivities

Description

Gets information about the scaling activities in the account and Region.

When scaling events occur, you see a record of the scaling activity in the scaling activities. For more information, see Verifying a scaling activity for an Auto Scaling group in the Amazon EC2 Auto Scaling User Guide.

If the scaling event succeeds, the value of the StatusCode element in the response is Successful. If an attempt to launch instances failed, the StatusCode value is Failed or Cancelled and the StatusMessage element in the response indicates the cause of the failure. For help interpreting the StatusMessage, see Troubleshooting Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

This operation returns paginated results.

Creating a Request

data DescribeScalingActivities

See: newDescribeScalingActivities smart constructor.

Constructors

DescribeScalingActivities'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

autoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

maxRecords :: Maybe Int

The maximum number of items to return with this call. The default value is 100 and the maximum value is 100.

includeDeletedGroups :: Maybe Bool

Indicates whether to include scaling activity from deleted Auto Scaling groups.

activityIds :: Maybe [Text]

The activity IDs of the desired scaling activities. If you omit this parameter, all activities for the past six weeks are described. If unknown activities are requested, they are ignored with no error. If you specify an Auto Scaling group, the results are limited to that group.

Array Members: Maximum number of 50 IDs.

newDescribeScalingActivities :: DescribeScalingActivities

Create a value of DescribeScalingActivities with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeScalingActivities', describeScalingActivities_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:autoScalingGroupName:DescribeScalingActivities', describeScalingActivities_autoScalingGroupName - The name of the Auto Scaling group.

$sel:maxRecords:DescribeScalingActivities', describeScalingActivities_maxRecords - The maximum number of items to return with this call. The default value is 100 and the maximum value is 100.

$sel:includeDeletedGroups:DescribeScalingActivities', describeScalingActivities_includeDeletedGroups - Indicates whether to include scaling activity from deleted Auto Scaling groups.

$sel:activityIds:DescribeScalingActivities', describeScalingActivities_activityIds - The activity IDs of the desired scaling activities. If you omit this parameter, all activities for the past six weeks are described. If unknown activities are requested, they are ignored with no error. If you specify an Auto Scaling group, the results are limited to that group.

Array Members: Maximum number of 50 IDs.

Request Lenses

describeScalingActivities_nextToken :: Lens' DescribeScalingActivities (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeScalingActivities_autoScalingGroupName :: Lens' DescribeScalingActivities (Maybe Text)

The name of the Auto Scaling group.

describeScalingActivities_maxRecords :: Lens' DescribeScalingActivities (Maybe Int)

The maximum number of items to return with this call. The default value is 100 and the maximum value is 100.

describeScalingActivities_includeDeletedGroups :: Lens' DescribeScalingActivities (Maybe Bool)

Indicates whether to include scaling activity from deleted Auto Scaling groups.

describeScalingActivities_activityIds :: Lens' DescribeScalingActivities (Maybe [Text])

The activity IDs of the desired scaling activities. If you omit this parameter, all activities for the past six weeks are described. If unknown activities are requested, they are ignored with no error. If you specify an Auto Scaling group, the results are limited to that group.

Array Members: Maximum number of 50 IDs.

Destructuring the Response

data DescribeScalingActivitiesResponse

See: newDescribeScalingActivitiesResponse smart constructor.

Constructors

DescribeScalingActivitiesResponse'

Fields

nextToken :: Maybe Text

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

httpStatus :: Int

The response's http status code.

activities :: [Activity]

The scaling activities. Activities are sorted by start time. Activities still in progress are described first.

newDescribeScalingActivitiesResponse

Arguments

:: Int
$sel:httpStatus:DescribeScalingActivitiesResponse'

-> DescribeScalingActivitiesResponse

Create a value of DescribeScalingActivitiesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeScalingActivities', describeScalingActivitiesResponse_nextToken - A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

$sel:httpStatus:DescribeScalingActivitiesResponse', describeScalingActivitiesResponse_httpStatus - The response's http status code.

$sel:activities:DescribeScalingActivitiesResponse', describeScalingActivitiesResponse_activities - The scaling activities. Activities are sorted by start time. Activities still in progress are described first.

Response Lenses

describeScalingActivitiesResponse_nextToken :: Lens' DescribeScalingActivitiesResponse (Maybe Text)

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

describeScalingActivitiesResponse_httpStatus :: Lens' DescribeScalingActivitiesResponse Int

The response's http status code.

describeScalingActivitiesResponse_activities :: Lens' DescribeScalingActivitiesResponse [Activity]

The scaling activities. Activities are sorted by start time. Activities still in progress are described first.