Network.AWS.AutoScaling.DescribeScheduledActions

Description

Gets information about the scheduled actions that haven't run or that have not reached their end time.

To describe the scaling activities for scheduled actions that have already run, call the DescribeScalingActivities API.

This operation returns paginated results.

Creating a Request

data DescribeScheduledActions

See: newDescribeScheduledActions smart constructor.

Constructors

DescribeScheduledActions'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

startTime :: Maybe ISO8601

The earliest scheduled start time to return. If scheduled action names are provided, this parameter is ignored.

endTime :: Maybe ISO8601

The latest scheduled start time to return. If scheduled action names are provided, this parameter is ignored.

scheduledActionNames :: Maybe [Text]

The names of one or more scheduled actions. If you omit this parameter, all scheduled actions are described. If you specify an unknown scheduled action, it is ignored with no error.

Array Members: Maximum number of 50 actions.

maxRecords :: Maybe Int

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

autoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

newDescribeScheduledActions :: DescribeScheduledActions

Create a value of DescribeScheduledActions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeScheduledActions', describeScheduledActions_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:startTime:DescribeScheduledActions', describeScheduledActions_startTime - The earliest scheduled start time to return. If scheduled action names are provided, this parameter is ignored.

$sel:endTime:DescribeScheduledActions', describeScheduledActions_endTime - The latest scheduled start time to return. If scheduled action names are provided, this parameter is ignored.

$sel:scheduledActionNames:DescribeScheduledActions', describeScheduledActions_scheduledActionNames - The names of one or more scheduled actions. If you omit this parameter, all scheduled actions are described. If you specify an unknown scheduled action, it is ignored with no error.

Array Members: Maximum number of 50 actions.

$sel:maxRecords:DescribeScheduledActions', describeScheduledActions_maxRecords - The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

$sel:autoScalingGroupName:DescribeScheduledActions', describeScheduledActions_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

describeScheduledActions_nextToken :: Lens' DescribeScheduledActions (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeScheduledActions_startTime :: Lens' DescribeScheduledActions (Maybe UTCTime)

The earliest scheduled start time to return. If scheduled action names are provided, this parameter is ignored.

describeScheduledActions_endTime :: Lens' DescribeScheduledActions (Maybe UTCTime)

The latest scheduled start time to return. If scheduled action names are provided, this parameter is ignored.

describeScheduledActions_scheduledActionNames :: Lens' DescribeScheduledActions (Maybe [Text])

The names of one or more scheduled actions. If you omit this parameter, all scheduled actions are described. If you specify an unknown scheduled action, it is ignored with no error.

Array Members: Maximum number of 50 actions.

describeScheduledActions_maxRecords :: Lens' DescribeScheduledActions (Maybe Int)

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

describeScheduledActions_autoScalingGroupName :: Lens' DescribeScheduledActions (Maybe Text)

The name of the Auto Scaling group.

Destructuring the Response

data DescribeScheduledActionsResponse

See: newDescribeScheduledActionsResponse smart constructor.

Constructors

DescribeScheduledActionsResponse'

Fields

nextToken :: Maybe Text

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

scheduledUpdateGroupActions :: Maybe [ScheduledUpdateGroupAction]

The scheduled actions.

httpStatus :: Int

The response's http status code.

newDescribeScheduledActionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeScheduledActionsResponse'

-> DescribeScheduledActionsResponse

Create a value of DescribeScheduledActionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeScheduledActions', describeScheduledActionsResponse_nextToken - A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

$sel:scheduledUpdateGroupActions:DescribeScheduledActionsResponse', describeScheduledActionsResponse_scheduledUpdateGroupActions - The scheduled actions.

$sel:httpStatus:DescribeScheduledActionsResponse', describeScheduledActionsResponse_httpStatus - The response's http status code.

Response Lenses

describeScheduledActionsResponse_nextToken :: Lens' DescribeScheduledActionsResponse (Maybe Text)

A string that indicates that the response contains more items than can be returned in a single response. To receive additional items, specify this string for the NextToken value when requesting the next set of items. This value is null when there are no more items to return.

describeScheduledActionsResponse_scheduledUpdateGroupActions :: Lens' DescribeScheduledActionsResponse (Maybe [ScheduledUpdateGroupAction])

The scheduled actions.

describeScheduledActionsResponse_httpStatus :: Lens' DescribeScheduledActionsResponse Int

The response's http status code.