Network.AWS.AutoScaling.DescribeWarmPool

Description

Gets information about a warm pool and its instances.

For more information, see Warm pools for Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

Creating a Request

data DescribeWarmPool

See: newDescribeWarmPool smart constructor.

Constructors

DescribeWarmPool'

Fields

nextToken :: Maybe Text

The token for the next set of instances to return. (You received this token from a previous call.)

maxRecords :: Maybe Int

The maximum number of instances to return with this call. The maximum value is 50.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

newDescribeWarmPool

Arguments

:: Text
$sel:autoScalingGroupName:DescribeWarmPool'

-> DescribeWarmPool

Create a value of DescribeWarmPool with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeWarmPool', describeWarmPool_nextToken - The token for the next set of instances to return. (You received this token from a previous call.)

$sel:maxRecords:DescribeWarmPool', describeWarmPool_maxRecords - The maximum number of instances to return with this call. The maximum value is 50.

$sel:autoScalingGroupName:DescribeWarmPool', describeWarmPool_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

describeWarmPool_nextToken :: Lens' DescribeWarmPool (Maybe Text)

The token for the next set of instances to return. (You received this token from a previous call.)

describeWarmPool_maxRecords :: Lens' DescribeWarmPool (Maybe Int)

The maximum number of instances to return with this call. The maximum value is 50.

describeWarmPool_autoScalingGroupName :: Lens' DescribeWarmPool Text

The name of the Auto Scaling group.

Destructuring the Response

data DescribeWarmPoolResponse

See: newDescribeWarmPoolResponse smart constructor.

Constructors

DescribeWarmPoolResponse'

Fields

nextToken :: Maybe Text

The token for the next set of items to return. (You received this token from a previous call.)

instances :: Maybe [Instance]

The instances that are currently in the warm pool.

warmPoolConfiguration :: Maybe WarmPoolConfiguration

The warm pool configuration details.

httpStatus :: Int

The response's http status code.

newDescribeWarmPoolResponse

Arguments

:: Int
$sel:httpStatus:DescribeWarmPoolResponse'

-> DescribeWarmPoolResponse

Create a value of DescribeWarmPoolResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeWarmPool', describeWarmPoolResponse_nextToken - The token for the next set of items to return. (You received this token from a previous call.)

$sel:instances:DescribeWarmPoolResponse', describeWarmPoolResponse_instances - The instances that are currently in the warm pool.

$sel:warmPoolConfiguration:DescribeWarmPoolResponse', describeWarmPoolResponse_warmPoolConfiguration - The warm pool configuration details.

$sel:httpStatus:DescribeWarmPoolResponse', describeWarmPoolResponse_httpStatus - The response's http status code.

Response Lenses

describeWarmPoolResponse_nextToken :: Lens' DescribeWarmPoolResponse (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

describeWarmPoolResponse_instances :: Lens' DescribeWarmPoolResponse (Maybe [Instance])

The instances that are currently in the warm pool.

describeWarmPoolResponse_warmPoolConfiguration :: Lens' DescribeWarmPoolResponse (Maybe WarmPoolConfiguration)

The warm pool configuration details.

describeWarmPoolResponse_httpStatus :: Lens' DescribeWarmPoolResponse Int

The response's http status code.