Network.AWS.AutoScaling.DetachLoadBalancerTargetGroups

Description

Detaches one or more target groups from the specified Auto Scaling group.

Creating a Request

data DetachLoadBalancerTargetGroups

See: newDetachLoadBalancerTargetGroups smart constructor.

Constructors

DetachLoadBalancerTargetGroups'

Fields

autoScalingGroupName :: Text

The name of the Auto Scaling group.

targetGroupARNs :: [Text]

The Amazon Resource Names (ARN) of the target groups. You can specify up to 10 target groups.

newDetachLoadBalancerTargetGroups

Arguments

:: Text
$sel:autoScalingGroupName:DetachLoadBalancerTargetGroups'

-> DetachLoadBalancerTargetGroups

Create a value of DetachLoadBalancerTargetGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoScalingGroupName:DetachLoadBalancerTargetGroups', detachLoadBalancerTargetGroups_autoScalingGroupName - The name of the Auto Scaling group.

$sel:targetGroupARNs:DetachLoadBalancerTargetGroups', detachLoadBalancerTargetGroups_targetGroupARNs - The Amazon Resource Names (ARN) of the target groups. You can specify up to 10 target groups.

Request Lenses

detachLoadBalancerTargetGroups_autoScalingGroupName :: Lens' DetachLoadBalancerTargetGroups Text

The name of the Auto Scaling group.

detachLoadBalancerTargetGroups_targetGroupARNs :: Lens' DetachLoadBalancerTargetGroups [Text]

The Amazon Resource Names (ARN) of the target groups. You can specify up to 10 target groups.

Destructuring the Response

data DetachLoadBalancerTargetGroupsResponse

See: newDetachLoadBalancerTargetGroupsResponse smart constructor.

Constructors

DetachLoadBalancerTargetGroupsResponse'

Fields

httpStatus :: Int

The response's http status code.

newDetachLoadBalancerTargetGroupsResponse

Arguments

:: Int
$sel:httpStatus:DetachLoadBalancerTargetGroupsResponse'

-> DetachLoadBalancerTargetGroupsResponse

Create a value of DetachLoadBalancerTargetGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DetachLoadBalancerTargetGroupsResponse', detachLoadBalancerTargetGroupsResponse_httpStatus - The response's http status code.

Response Lenses

detachLoadBalancerTargetGroupsResponse_httpStatus :: Lens' DetachLoadBalancerTargetGroupsResponse Int

The response's http status code.