Network.AWS.AutoScaling.DetachLoadBalancers

Description

Detaches one or more Classic Load Balancers from the specified Auto Scaling group.

This operation detaches only Classic Load Balancers. If you have Application Load Balancers, Network Load Balancers, or Gateway Load Balancers, use the DetachLoadBalancerTargetGroups API instead.

When you detach a load balancer, it enters the Removing state while deregistering the instances in the group. When all instances are deregistered, then you can no longer describe the load balancer using the DescribeLoadBalancers API call. The instances remain running.

Creating a Request

data DetachLoadBalancers

See: newDetachLoadBalancers smart constructor.

Constructors

DetachLoadBalancers'

Fields

autoScalingGroupName :: Text

The name of the Auto Scaling group.

loadBalancerNames :: [Text]

The names of the load balancers. You can specify up to 10 load balancers.

newDetachLoadBalancers

Arguments

:: Text
$sel:autoScalingGroupName:DetachLoadBalancers'

-> DetachLoadBalancers

Create a value of DetachLoadBalancers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoScalingGroupName:DetachLoadBalancers', detachLoadBalancers_autoScalingGroupName - The name of the Auto Scaling group.

$sel:loadBalancerNames:DetachLoadBalancers', detachLoadBalancers_loadBalancerNames - The names of the load balancers. You can specify up to 10 load balancers.

Request Lenses

detachLoadBalancers_autoScalingGroupName :: Lens' DetachLoadBalancers Text

The name of the Auto Scaling group.

detachLoadBalancers_loadBalancerNames :: Lens' DetachLoadBalancers [Text]

The names of the load balancers. You can specify up to 10 load balancers.

Destructuring the Response

data DetachLoadBalancersResponse

See: newDetachLoadBalancersResponse smart constructor.

Constructors

DetachLoadBalancersResponse'

Fields

httpStatus :: Int

The response's http status code.

newDetachLoadBalancersResponse

Arguments

:: Int
$sel:httpStatus:DetachLoadBalancersResponse'

-> DetachLoadBalancersResponse

Create a value of DetachLoadBalancersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DetachLoadBalancersResponse', detachLoadBalancersResponse_httpStatus - The response's http status code.

Response Lenses

detachLoadBalancersResponse_httpStatus :: Lens' DetachLoadBalancersResponse Int

The response's http status code.