Network.AWS.AutoScaling.ExecutePolicy

Description

Executes the specified policy. This can be useful for testing the design of your scaling policy.

Creating a Request

data ExecutePolicy

See: newExecutePolicy smart constructor.

Constructors

ExecutePolicy'

Fields

metricValue :: Maybe Double

The metric value to compare to BreachThreshold. This enables you to execute a policy of type StepScaling and determine which step adjustment to use. For example, if the breach threshold is 50 and you want to use a step adjustment with a lower bound of 0 and an upper bound of 10, you can set the metric value to 59.

If you specify a metric value that doesn't correspond to a step adjustment for the policy, the call returns an error.

Required if the policy type is StepScaling and not supported otherwise.

breachThreshold :: Maybe Double

The breach threshold for the alarm.

Required if the policy type is StepScaling and not supported otherwise.

honorCooldown :: Maybe Bool

Indicates whether Amazon EC2 Auto Scaling waits for the cooldown period to complete before executing the policy.

Valid only if the policy type is SimpleScaling. For more information, see Scaling cooldowns for Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

autoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

policyName :: Text

The name or ARN of the policy.

newExecutePolicy

Arguments

:: Text
$sel:policyName:ExecutePolicy'

-> ExecutePolicy

Create a value of ExecutePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricValue:ExecutePolicy', executePolicy_metricValue - The metric value to compare to BreachThreshold. This enables you to execute a policy of type StepScaling and determine which step adjustment to use. For example, if the breach threshold is 50 and you want to use a step adjustment with a lower bound of 0 and an upper bound of 10, you can set the metric value to 59.

If you specify a metric value that doesn't correspond to a step adjustment for the policy, the call returns an error.

Required if the policy type is StepScaling and not supported otherwise.

$sel:breachThreshold:ExecutePolicy', executePolicy_breachThreshold - The breach threshold for the alarm.

Required if the policy type is StepScaling and not supported otherwise.

$sel:honorCooldown:ExecutePolicy', executePolicy_honorCooldown - Indicates whether Amazon EC2 Auto Scaling waits for the cooldown period to complete before executing the policy.

Valid only if the policy type is SimpleScaling. For more information, see Scaling cooldowns for Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

$sel:autoScalingGroupName:ExecutePolicy', executePolicy_autoScalingGroupName - The name of the Auto Scaling group.

$sel:policyName:ExecutePolicy', executePolicy_policyName - The name or ARN of the policy.

Request Lenses

executePolicy_metricValue :: Lens' ExecutePolicy (Maybe Double)

The metric value to compare to BreachThreshold. This enables you to execute a policy of type StepScaling and determine which step adjustment to use. For example, if the breach threshold is 50 and you want to use a step adjustment with a lower bound of 0 and an upper bound of 10, you can set the metric value to 59.

If you specify a metric value that doesn't correspond to a step adjustment for the policy, the call returns an error.

Required if the policy type is StepScaling and not supported otherwise.

executePolicy_breachThreshold :: Lens' ExecutePolicy (Maybe Double)

The breach threshold for the alarm.

Required if the policy type is StepScaling and not supported otherwise.

executePolicy_honorCooldown :: Lens' ExecutePolicy (Maybe Bool)

Indicates whether Amazon EC2 Auto Scaling waits for the cooldown period to complete before executing the policy.

Valid only if the policy type is SimpleScaling. For more information, see Scaling cooldowns for Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

executePolicy_autoScalingGroupName :: Lens' ExecutePolicy (Maybe Text)

The name of the Auto Scaling group.

executePolicy_policyName :: Lens' ExecutePolicy Text

The name or ARN of the policy.

Destructuring the Response

data ExecutePolicyResponse

See: newExecutePolicyResponse smart constructor.

Constructors

ExecutePolicyResponse'

newExecutePolicyResponse :: ExecutePolicyResponse

Create a value of ExecutePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.