Network.AWS.AutoScaling.GetPredictiveScalingForecast

Description

Retrieves the forecast data for a predictive scaling policy.

Load forecasts are predictions of the hourly load values using historical load data from CloudWatch and an analysis of historical trends. Capacity forecasts are represented as predicted values for the minimum capacity that is needed on an hourly basis, based on the hourly load forecast.

A minimum of 24 hours of data is required to create the initial forecasts. However, having a full 14 days of historical data results in more accurate forecasts.

For more information, see Predictive scaling for Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

Creating a Request

data GetPredictiveScalingForecast

See: newGetPredictiveScalingForecast smart constructor.

Constructors

GetPredictiveScalingForecast'

Fields

autoScalingGroupName :: Text

The name of the Auto Scaling group.

policyName :: Text

The name of the policy.

startTime :: ISO8601

The inclusive start time of the time range for the forecast data to get. At most, the date and time can be one year before the current date and time.

endTime :: ISO8601

The exclusive end time of the time range for the forecast data to get. The maximum time duration between the start and end time is 30 days.

Although this parameter can accept a date and time that is more than two days in the future, the availability of forecast data has limits. Amazon EC2 Auto Scaling only issues forecasts for periods of two days in advance.

newGetPredictiveScalingForecast

Arguments

:: Text
$sel:autoScalingGroupName:GetPredictiveScalingForecast'

-> Text
$sel:policyName:GetPredictiveScalingForecast'

-> UTCTime
$sel:startTime:GetPredictiveScalingForecast'

-> UTCTime
$sel:endTime:GetPredictiveScalingForecast'

-> GetPredictiveScalingForecast

Create a value of GetPredictiveScalingForecast with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoScalingGroupName:GetPredictiveScalingForecast', getPredictiveScalingForecast_autoScalingGroupName - The name of the Auto Scaling group.

$sel:policyName:GetPredictiveScalingForecast', getPredictiveScalingForecast_policyName - The name of the policy.

$sel:startTime:GetPredictiveScalingForecast', getPredictiveScalingForecast_startTime - The inclusive start time of the time range for the forecast data to get. At most, the date and time can be one year before the current date and time.

$sel:endTime:GetPredictiveScalingForecast', getPredictiveScalingForecast_endTime - The exclusive end time of the time range for the forecast data to get. The maximum time duration between the start and end time is 30 days.

Although this parameter can accept a date and time that is more than two days in the future, the availability of forecast data has limits. Amazon EC2 Auto Scaling only issues forecasts for periods of two days in advance.

Request Lenses

getPredictiveScalingForecast_autoScalingGroupName :: Lens' GetPredictiveScalingForecast Text

The name of the Auto Scaling group.

getPredictiveScalingForecast_policyName :: Lens' GetPredictiveScalingForecast Text

The name of the policy.

getPredictiveScalingForecast_startTime :: Lens' GetPredictiveScalingForecast UTCTime

The inclusive start time of the time range for the forecast data to get. At most, the date and time can be one year before the current date and time.

getPredictiveScalingForecast_endTime :: Lens' GetPredictiveScalingForecast UTCTime

The exclusive end time of the time range for the forecast data to get. The maximum time duration between the start and end time is 30 days.

Although this parameter can accept a date and time that is more than two days in the future, the availability of forecast data has limits. Amazon EC2 Auto Scaling only issues forecasts for periods of two days in advance.

Destructuring the Response

data GetPredictiveScalingForecastResponse

See: newGetPredictiveScalingForecastResponse smart constructor.

Constructors

GetPredictiveScalingForecastResponse'

Fields

httpStatus :: Int

The response's http status code.

loadForecast :: [LoadForecast]

The load forecast.

capacityForecast :: CapacityForecast

The capacity forecast.

updateTime :: ISO8601

The time the forecast was made.

newGetPredictiveScalingForecastResponse

Arguments

:: Int
$sel:httpStatus:GetPredictiveScalingForecastResponse'

-> CapacityForecast
$sel:capacityForecast:GetPredictiveScalingForecastResponse'

-> UTCTime
$sel:updateTime:GetPredictiveScalingForecastResponse'

-> GetPredictiveScalingForecastResponse

Create a value of GetPredictiveScalingForecastResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetPredictiveScalingForecastResponse', getPredictiveScalingForecastResponse_httpStatus - The response's http status code.

$sel:loadForecast:GetPredictiveScalingForecastResponse', getPredictiveScalingForecastResponse_loadForecast - The load forecast.

$sel:capacityForecast:GetPredictiveScalingForecastResponse', getPredictiveScalingForecastResponse_capacityForecast - The capacity forecast.

$sel:updateTime:GetPredictiveScalingForecastResponse', getPredictiveScalingForecastResponse_updateTime - The time the forecast was made.

Response Lenses

getPredictiveScalingForecastResponse_httpStatus :: Lens' GetPredictiveScalingForecastResponse Int

The response's http status code.

getPredictiveScalingForecastResponse_loadForecast :: Lens' GetPredictiveScalingForecastResponse [LoadForecast]

The load forecast.

getPredictiveScalingForecastResponse_capacityForecast :: Lens' GetPredictiveScalingForecastResponse CapacityForecast

The capacity forecast.

getPredictiveScalingForecastResponse_updateTime :: Lens' GetPredictiveScalingForecastResponse UTCTime

The time the forecast was made.