Network.AWS.AutoScaling.PutScalingPolicy

Description

Creates or updates a scaling policy for an Auto Scaling group. Scaling policies are used to scale an Auto Scaling group based on configurable metrics. If no policies are defined, the dynamic scaling and predictive scaling features are not used.

For more information about using dynamic scaling, see Target tracking scaling policies and Step and simple scaling policies in the Amazon EC2 Auto Scaling User Guide.

For more information about using predictive scaling, see Predictive scaling for Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

You can view the scaling policies for an Auto Scaling group using the DescribePolicies API call. If you are no longer using a scaling policy, you can delete it by calling the DeletePolicy API.

Creating a Request

data PutScalingPolicy

See: newPutScalingPolicy smart constructor.

Constructors

PutScalingPolicy'

Fields

minAdjustmentStep :: Maybe Int

Available for backward compatibility. Use MinAdjustmentMagnitude instead.

estimatedInstanceWarmup :: Maybe Int

The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics. If not provided, the default is to use the value from the default cooldown period for the Auto Scaling group.

Valid only if the policy type is TargetTrackingScaling or StepScaling.

enabled :: Maybe Bool

Indicates whether the scaling policy is enabled or disabled. The default is enabled. For more information, see Disabling a scaling policy for an Auto Scaling group in the Amazon EC2 Auto Scaling User Guide.

policyType :: Maybe Text

One of the following policy types:

TargetTrackingScaling

StepScaling

SimpleScaling (default)

PredictiveScaling

stepAdjustments :: Maybe [StepAdjustment]

A set of adjustments that enable you to scale based on the size of the alarm breach.

Required if the policy type is StepScaling. (Not used with any other policy type.)

targetTrackingConfiguration :: Maybe TargetTrackingConfiguration

A target tracking scaling policy. Provides support for predefined or customized metrics.

The following predefined metrics are available:

ASGAverageCPUUtilization

ASGAverageNetworkIn

ASGAverageNetworkOut

ALBRequestCountPerTarget

If you specify ALBRequestCountPerTarget for the metric, you must specify the ResourceLabel parameter with the PredefinedMetricSpecification.

For more information, see TargetTrackingConfiguration in the Amazon EC2 Auto Scaling API Reference.

Required if the policy type is TargetTrackingScaling.

adjustmentType :: Maybe Text

Specifies how the scaling adjustment is interpreted (for example, an absolute number or a percentage). The valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

Required if the policy type is StepScaling or SimpleScaling. For more information, see Scaling adjustment types in the Amazon EC2 Auto Scaling User Guide.

predictiveScalingConfiguration :: Maybe PredictiveScalingConfiguration

A predictive scaling policy. Provides support for only predefined metrics.

Predictive scaling works with CPU utilization, network in/out, and the Application Load Balancer request count.

For more information, see PredictiveScalingConfiguration in the Amazon EC2 Auto Scaling API Reference.

Required if the policy type is PredictiveScaling.

scalingAdjustment :: Maybe Int

The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity. For exact capacity, you must specify a positive value.

Required if the policy type is SimpleScaling. (Not used with any other policy type.)

cooldown :: Maybe Int

The duration of the policy's cooldown period, in seconds. When a cooldown period is specified here, it overrides the default cooldown period defined for the Auto Scaling group.

Valid only if the policy type is SimpleScaling. For more information, see Scaling cooldowns for Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

metricAggregationType :: Maybe Text

The aggregation type for the CloudWatch metrics. The valid values are Minimum, Maximum, and Average. If the aggregation type is null, the value is treated as Average.

Valid only if the policy type is StepScaling.

minAdjustmentMagnitude :: Maybe Int

The minimum value to scale by when the adjustment type is PercentChangeInCapacity. For example, suppose that you create a step scaling policy to scale out an Auto Scaling group by 25 percent and you specify a MinAdjustmentMagnitude of 2. If the group has 4 instances and the scaling policy is performed, 25 percent of 4 is 1. However, because you specified a MinAdjustmentMagnitude of 2, Amazon EC2 Auto Scaling scales out the group by 2 instances.

Valid only if the policy type is StepScaling or SimpleScaling. For more information, see Scaling adjustment types in the Amazon EC2 Auto Scaling User Guide.

Some Auto Scaling groups use instance weights. In this case, set the MinAdjustmentMagnitude to a value that is at least as large as your largest instance weight.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

policyName :: Text

The name of the policy.

newPutScalingPolicy

Arguments

:: Text
$sel:autoScalingGroupName:PutScalingPolicy'

-> Text
$sel:policyName:PutScalingPolicy'

-> PutScalingPolicy

Create a value of PutScalingPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minAdjustmentStep:PutScalingPolicy', putScalingPolicy_minAdjustmentStep - Available for backward compatibility. Use MinAdjustmentMagnitude instead.

$sel:estimatedInstanceWarmup:PutScalingPolicy', putScalingPolicy_estimatedInstanceWarmup - The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics. If not provided, the default is to use the value from the default cooldown period for the Auto Scaling group.

Valid only if the policy type is TargetTrackingScaling or StepScaling.

$sel:enabled:PutScalingPolicy', putScalingPolicy_enabled - Indicates whether the scaling policy is enabled or disabled. The default is enabled. For more information, see Disabling a scaling policy for an Auto Scaling group in the Amazon EC2 Auto Scaling User Guide.

$sel:policyType:PutScalingPolicy', putScalingPolicy_policyType - One of the following policy types:

TargetTrackingScaling

StepScaling

SimpleScaling (default)

PredictiveScaling

$sel:stepAdjustments:PutScalingPolicy', putScalingPolicy_stepAdjustments - A set of adjustments that enable you to scale based on the size of the alarm breach.

Required if the policy type is StepScaling. (Not used with any other policy type.)

$sel:targetTrackingConfiguration:PutScalingPolicy', putScalingPolicy_targetTrackingConfiguration - A target tracking scaling policy. Provides support for predefined or customized metrics.

The following predefined metrics are available:

ASGAverageCPUUtilization

ASGAverageNetworkIn

ASGAverageNetworkOut

ALBRequestCountPerTarget

If you specify ALBRequestCountPerTarget for the metric, you must specify the ResourceLabel parameter with the PredefinedMetricSpecification.

For more information, see TargetTrackingConfiguration in the Amazon EC2 Auto Scaling API Reference.

Required if the policy type is TargetTrackingScaling.

$sel:adjustmentType:PutScalingPolicy', putScalingPolicy_adjustmentType - Specifies how the scaling adjustment is interpreted (for example, an absolute number or a percentage). The valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

Required if the policy type is StepScaling or SimpleScaling. For more information, see Scaling adjustment types in the Amazon EC2 Auto Scaling User Guide.

$sel:predictiveScalingConfiguration:PutScalingPolicy', putScalingPolicy_predictiveScalingConfiguration - A predictive scaling policy. Provides support for only predefined metrics.

Predictive scaling works with CPU utilization, network in/out, and the Application Load Balancer request count.

For more information, see PredictiveScalingConfiguration in the Amazon EC2 Auto Scaling API Reference.

Required if the policy type is PredictiveScaling.

$sel:scalingAdjustment:PutScalingPolicy', putScalingPolicy_scalingAdjustment - The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity. For exact capacity, you must specify a positive value.

Required if the policy type is SimpleScaling. (Not used with any other policy type.)

$sel:cooldown:PutScalingPolicy', putScalingPolicy_cooldown - The duration of the policy's cooldown period, in seconds. When a cooldown period is specified here, it overrides the default cooldown period defined for the Auto Scaling group.

Valid only if the policy type is SimpleScaling. For more information, see Scaling cooldowns for Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

$sel:metricAggregationType:PutScalingPolicy', putScalingPolicy_metricAggregationType - The aggregation type for the CloudWatch metrics. The valid values are Minimum, Maximum, and Average. If the aggregation type is null, the value is treated as Average.

Valid only if the policy type is StepScaling.

$sel:minAdjustmentMagnitude:PutScalingPolicy', putScalingPolicy_minAdjustmentMagnitude - The minimum value to scale by when the adjustment type is PercentChangeInCapacity. For example, suppose that you create a step scaling policy to scale out an Auto Scaling group by 25 percent and you specify a MinAdjustmentMagnitude of 2. If the group has 4 instances and the scaling policy is performed, 25 percent of 4 is 1. However, because you specified a MinAdjustmentMagnitude of 2, Amazon EC2 Auto Scaling scales out the group by 2 instances.

Valid only if the policy type is StepScaling or SimpleScaling. For more information, see Scaling adjustment types in the Amazon EC2 Auto Scaling User Guide.

Some Auto Scaling groups use instance weights. In this case, set the MinAdjustmentMagnitude to a value that is at least as large as your largest instance weight.

$sel:autoScalingGroupName:PutScalingPolicy', putScalingPolicy_autoScalingGroupName - The name of the Auto Scaling group.

$sel:policyName:PutScalingPolicy', putScalingPolicy_policyName - The name of the policy.

Request Lenses

putScalingPolicy_minAdjustmentStep :: Lens' PutScalingPolicy (Maybe Int)

Available for backward compatibility. Use MinAdjustmentMagnitude instead.

putScalingPolicy_estimatedInstanceWarmup :: Lens' PutScalingPolicy (Maybe Int)

The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics. If not provided, the default is to use the value from the default cooldown period for the Auto Scaling group.

Valid only if the policy type is TargetTrackingScaling or StepScaling.

putScalingPolicy_enabled :: Lens' PutScalingPolicy (Maybe Bool)

Indicates whether the scaling policy is enabled or disabled. The default is enabled. For more information, see Disabling a scaling policy for an Auto Scaling group in the Amazon EC2 Auto Scaling User Guide.

putScalingPolicy_policyType :: Lens' PutScalingPolicy (Maybe Text)

One of the following policy types:

TargetTrackingScaling

StepScaling

SimpleScaling (default)

PredictiveScaling

putScalingPolicy_stepAdjustments :: Lens' PutScalingPolicy (Maybe [StepAdjustment])

A set of adjustments that enable you to scale based on the size of the alarm breach.

Required if the policy type is StepScaling. (Not used with any other policy type.)

putScalingPolicy_targetTrackingConfiguration :: Lens' PutScalingPolicy (Maybe TargetTrackingConfiguration)

A target tracking scaling policy. Provides support for predefined or customized metrics.

The following predefined metrics are available:

ASGAverageCPUUtilization

ASGAverageNetworkIn

ASGAverageNetworkOut

ALBRequestCountPerTarget

If you specify ALBRequestCountPerTarget for the metric, you must specify the ResourceLabel parameter with the PredefinedMetricSpecification.

For more information, see TargetTrackingConfiguration in the Amazon EC2 Auto Scaling API Reference.

Required if the policy type is TargetTrackingScaling.

putScalingPolicy_adjustmentType :: Lens' PutScalingPolicy (Maybe Text)

Specifies how the scaling adjustment is interpreted (for example, an absolute number or a percentage). The valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

Required if the policy type is StepScaling or SimpleScaling. For more information, see Scaling adjustment types in the Amazon EC2 Auto Scaling User Guide.

putScalingPolicy_predictiveScalingConfiguration :: Lens' PutScalingPolicy (Maybe PredictiveScalingConfiguration)

A predictive scaling policy. Provides support for only predefined metrics.

Predictive scaling works with CPU utilization, network in/out, and the Application Load Balancer request count.

For more information, see PredictiveScalingConfiguration in the Amazon EC2 Auto Scaling API Reference.

Required if the policy type is PredictiveScaling.

putScalingPolicy_scalingAdjustment :: Lens' PutScalingPolicy (Maybe Int)

The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity. For exact capacity, you must specify a positive value.

Required if the policy type is SimpleScaling. (Not used with any other policy type.)

putScalingPolicy_cooldown :: Lens' PutScalingPolicy (Maybe Int)

The duration of the policy's cooldown period, in seconds. When a cooldown period is specified here, it overrides the default cooldown period defined for the Auto Scaling group.

Valid only if the policy type is SimpleScaling. For more information, see Scaling cooldowns for Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

putScalingPolicy_metricAggregationType :: Lens' PutScalingPolicy (Maybe Text)

The aggregation type for the CloudWatch metrics. The valid values are Minimum, Maximum, and Average. If the aggregation type is null, the value is treated as Average.

Valid only if the policy type is StepScaling.

putScalingPolicy_minAdjustmentMagnitude :: Lens' PutScalingPolicy (Maybe Int)

The minimum value to scale by when the adjustment type is PercentChangeInCapacity. For example, suppose that you create a step scaling policy to scale out an Auto Scaling group by 25 percent and you specify a MinAdjustmentMagnitude of 2. If the group has 4 instances and the scaling policy is performed, 25 percent of 4 is 1. However, because you specified a MinAdjustmentMagnitude of 2, Amazon EC2 Auto Scaling scales out the group by 2 instances.

Valid only if the policy type is StepScaling or SimpleScaling. For more information, see Scaling adjustment types in the Amazon EC2 Auto Scaling User Guide.

Some Auto Scaling groups use instance weights. In this case, set the MinAdjustmentMagnitude to a value that is at least as large as your largest instance weight.

putScalingPolicy_autoScalingGroupName :: Lens' PutScalingPolicy Text

The name of the Auto Scaling group.

putScalingPolicy_policyName :: Lens' PutScalingPolicy Text

The name of the policy.

Destructuring the Response

data PutScalingPolicyResponse

Contains the output of PutScalingPolicy.

See: newPutScalingPolicyResponse smart constructor.

Constructors

PutScalingPolicyResponse'

Fields

policyARN :: Maybe Text

The Amazon Resource Name (ARN) of the policy.

alarms :: Maybe [Alarm]

The CloudWatch alarms created for the target tracking scaling policy.

httpStatus :: Int

The response's http status code.

newPutScalingPolicyResponse

Arguments

:: Int
$sel:httpStatus:PutScalingPolicyResponse'

-> PutScalingPolicyResponse

Create a value of PutScalingPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyARN:PutScalingPolicyResponse', putScalingPolicyResponse_policyARN - The Amazon Resource Name (ARN) of the policy.

$sel:alarms:PutScalingPolicyResponse', putScalingPolicyResponse_alarms - The CloudWatch alarms created for the target tracking scaling policy.

$sel:httpStatus:PutScalingPolicyResponse', putScalingPolicyResponse_httpStatus - The response's http status code.

Response Lenses

putScalingPolicyResponse_policyARN :: Lens' PutScalingPolicyResponse (Maybe Text)

The Amazon Resource Name (ARN) of the policy.

putScalingPolicyResponse_alarms :: Lens' PutScalingPolicyResponse (Maybe [Alarm])

The CloudWatch alarms created for the target tracking scaling policy.

putScalingPolicyResponse_httpStatus :: Lens' PutScalingPolicyResponse Int

The response's http status code.