Network.AWS.AutoScaling.PutScheduledUpdateGroupAction

Description

Creates or updates a scheduled scaling action for an Auto Scaling group.

For more information, see Scheduled scaling in the Amazon EC2 Auto Scaling User Guide.

You can view the scheduled actions for an Auto Scaling group using the DescribeScheduledActions API call. If you are no longer using a scheduled action, you can delete it by calling the DeleteScheduledAction API.

Creating a Request

data PutScheduledUpdateGroupAction

See: newPutScheduledUpdateGroupAction smart constructor.

Constructors

PutScheduledUpdateGroupAction'

Fields

minSize :: Maybe Int

The minimum size of the Auto Scaling group.

desiredCapacity :: Maybe Int

The desired capacity is the initial capacity of the Auto Scaling group after the scheduled action runs and the capacity it attempts to maintain. It can scale beyond this capacity if you add more scaling conditions.

startTime :: Maybe ISO8601

The date and time for this action to start, in YYYY-MM-DDThh:mm:ssZ format in UTC/GMT only and in quotes (for example, "2019-06-01T00:00:00Z").

If you specify Recurrence and StartTime, Amazon EC2 Auto Scaling performs the action at this time, and then performs the action based on the specified recurrence.

If you try to schedule your action in the past, Amazon EC2 Auto Scaling returns an error message.

endTime :: Maybe ISO8601

The date and time for the recurring schedule to end, in UTC.

recurrence :: Maybe Text

The recurring schedule for this action. This format consists of five fields separated by white spaces: [Minute] [Hour] [Day_of_Month] [Month_of_Year] [Day_of_Week]. The value must be in quotes (for example, "30 0 1 1,6,12 *"). For more information about this format, see Crontab.

When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action starts and stops.

Cron expressions use Universal Coordinated Time (UTC) by default.

maxSize :: Maybe Int

The maximum size of the Auto Scaling group.

timeZone :: Maybe Text

Specifies the time zone for a cron expression. If a time zone is not provided, UTC is used by default.

Valid values are the canonical names of the IANA time zones, derived from the IANA Time Zone Database (such as Etc/GMT+9 or Pacific/Tahiti). For more information, see https://en.wikipedia.org/wiki/List_of_tz_database_time_zones.

time :: Maybe ISO8601

This parameter is no longer used.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

scheduledActionName :: Text

The name of this scaling action.

newPutScheduledUpdateGroupAction

Arguments

:: Text
$sel:autoScalingGroupName:PutScheduledUpdateGroupAction'

-> Text
$sel:scheduledActionName:PutScheduledUpdateGroupAction'

-> PutScheduledUpdateGroupAction

Create a value of PutScheduledUpdateGroupAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minSize:PutScheduledUpdateGroupAction', putScheduledUpdateGroupAction_minSize - The minimum size of the Auto Scaling group.

$sel:desiredCapacity:PutScheduledUpdateGroupAction', putScheduledUpdateGroupAction_desiredCapacity - The desired capacity is the initial capacity of the Auto Scaling group after the scheduled action runs and the capacity it attempts to maintain. It can scale beyond this capacity if you add more scaling conditions.

$sel:startTime:PutScheduledUpdateGroupAction', putScheduledUpdateGroupAction_startTime - The date and time for this action to start, in YYYY-MM-DDThh:mm:ssZ format in UTC/GMT only and in quotes (for example, "2019-06-01T00:00:00Z").

If you specify Recurrence and StartTime, Amazon EC2 Auto Scaling performs the action at this time, and then performs the action based on the specified recurrence.

If you try to schedule your action in the past, Amazon EC2 Auto Scaling returns an error message.

$sel:endTime:PutScheduledUpdateGroupAction', putScheduledUpdateGroupAction_endTime - The date and time for the recurring schedule to end, in UTC.

$sel:recurrence:PutScheduledUpdateGroupAction', putScheduledUpdateGroupAction_recurrence - The recurring schedule for this action. This format consists of five fields separated by white spaces: [Minute] [Hour] [Day_of_Month] [Month_of_Year] [Day_of_Week]. The value must be in quotes (for example, "30 0 1 1,6,12 *"). For more information about this format, see Crontab.

When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action starts and stops.

Cron expressions use Universal Coordinated Time (UTC) by default.

$sel:maxSize:PutScheduledUpdateGroupAction', putScheduledUpdateGroupAction_maxSize - The maximum size of the Auto Scaling group.

$sel:timeZone:PutScheduledUpdateGroupAction', putScheduledUpdateGroupAction_timeZone - Specifies the time zone for a cron expression. If a time zone is not provided, UTC is used by default.

Valid values are the canonical names of the IANA time zones, derived from the IANA Time Zone Database (such as Etc/GMT+9 or Pacific/Tahiti). For more information, see https://en.wikipedia.org/wiki/List_of_tz_database_time_zones.

$sel:time:PutScheduledUpdateGroupAction', putScheduledUpdateGroupAction_time - This parameter is no longer used.

$sel:autoScalingGroupName:PutScheduledUpdateGroupAction', putScheduledUpdateGroupAction_autoScalingGroupName - The name of the Auto Scaling group.

$sel:scheduledActionName:PutScheduledUpdateGroupAction', putScheduledUpdateGroupAction_scheduledActionName - The name of this scaling action.

Request Lenses

putScheduledUpdateGroupAction_minSize :: Lens' PutScheduledUpdateGroupAction (Maybe Int)

The minimum size of the Auto Scaling group.

putScheduledUpdateGroupAction_desiredCapacity :: Lens' PutScheduledUpdateGroupAction (Maybe Int)

The desired capacity is the initial capacity of the Auto Scaling group after the scheduled action runs and the capacity it attempts to maintain. It can scale beyond this capacity if you add more scaling conditions.

putScheduledUpdateGroupAction_startTime :: Lens' PutScheduledUpdateGroupAction (Maybe UTCTime)

The date and time for this action to start, in YYYY-MM-DDThh:mm:ssZ format in UTC/GMT only and in quotes (for example, "2019-06-01T00:00:00Z").

If you specify Recurrence and StartTime, Amazon EC2 Auto Scaling performs the action at this time, and then performs the action based on the specified recurrence.

If you try to schedule your action in the past, Amazon EC2 Auto Scaling returns an error message.

putScheduledUpdateGroupAction_endTime :: Lens' PutScheduledUpdateGroupAction (Maybe UTCTime)

The date and time for the recurring schedule to end, in UTC.

putScheduledUpdateGroupAction_recurrence :: Lens' PutScheduledUpdateGroupAction (Maybe Text)

The recurring schedule for this action. This format consists of five fields separated by white spaces: [Minute] [Hour] [Day_of_Month] [Month_of_Year] [Day_of_Week]. The value must be in quotes (for example, "30 0 1 1,6,12 *"). For more information about this format, see Crontab.

When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action starts and stops.

Cron expressions use Universal Coordinated Time (UTC) by default.

putScheduledUpdateGroupAction_maxSize :: Lens' PutScheduledUpdateGroupAction (Maybe Int)

The maximum size of the Auto Scaling group.

putScheduledUpdateGroupAction_timeZone :: Lens' PutScheduledUpdateGroupAction (Maybe Text)

Specifies the time zone for a cron expression. If a time zone is not provided, UTC is used by default.

Valid values are the canonical names of the IANA time zones, derived from the IANA Time Zone Database (such as Etc/GMT+9 or Pacific/Tahiti). For more information, see https://en.wikipedia.org/wiki/List_of_tz_database_time_zones.

putScheduledUpdateGroupAction_time :: Lens' PutScheduledUpdateGroupAction (Maybe UTCTime)

This parameter is no longer used.

putScheduledUpdateGroupAction_autoScalingGroupName :: Lens' PutScheduledUpdateGroupAction Text

The name of the Auto Scaling group.

putScheduledUpdateGroupAction_scheduledActionName :: Lens' PutScheduledUpdateGroupAction Text

The name of this scaling action.

Destructuring the Response

data PutScheduledUpdateGroupActionResponse

See: newPutScheduledUpdateGroupActionResponse smart constructor.

Constructors

PutScheduledUpdateGroupActionResponse'

newPutScheduledUpdateGroupActionResponse :: PutScheduledUpdateGroupActionResponse

Create a value of PutScheduledUpdateGroupActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.