Network.AWS.AutoScaling.PutWarmPool

Description

Creates or updates a warm pool for the specified Auto Scaling group. A warm pool is a pool of pre-initialized EC2 instances that sits alongside the Auto Scaling group. Whenever your application needs to scale out, the Auto Scaling group can draw on the warm pool to meet its new desired capacity. For more information and example configurations, see Warm pools for Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

This operation must be called from the Region in which the Auto Scaling group was created. This operation cannot be called on an Auto Scaling group that has a mixed instances policy or a launch template or launch configuration that requests Spot Instances.

You can view the instances in the warm pool using the DescribeWarmPool API call. If you are no longer using a warm pool, you can delete it by calling the DeleteWarmPool API.

Creating a Request

data PutWarmPool

See: newPutWarmPool smart constructor.

Constructors

PutWarmPool'

Fields

minSize :: Maybe Natural

Specifies the minimum number of instances to maintain in the warm pool. This helps you to ensure that there is always a certain number of warmed instances available to handle traffic spikes. Defaults to 0 if not specified.

maxGroupPreparedCapacity :: Maybe Int

Specifies the maximum number of instances that are allowed to be in the warm pool or in any state except Terminated for the Auto Scaling group. This is an optional property. Specify it only if you do not want the warm pool size to be determined by the difference between the group's maximum capacity and its desired capacity.

If a value for MaxGroupPreparedCapacity is not specified, Amazon EC2 Auto Scaling launches and maintains the difference between the group's maximum capacity and its desired capacity. If you specify a value for MaxGroupPreparedCapacity, Amazon EC2 Auto Scaling uses the difference between the MaxGroupPreparedCapacity and the desired capacity instead.

The size of the warm pool is dynamic. Only when MaxGroupPreparedCapacity and MinSize are set to the same value does the warm pool have an absolute size.

If the desired capacity of the Auto Scaling group is higher than the MaxGroupPreparedCapacity, the capacity of the warm pool is 0, unless you specify a value for MinSize. To remove a value that you previously set, include the property but specify -1 for the value.

poolState :: Maybe WarmPoolState

Sets the instance state to transition to after the lifecycle actions are complete. Default is Stopped.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

newPutWarmPool

Arguments

:: Text
$sel:autoScalingGroupName:PutWarmPool'

-> PutWarmPool

Create a value of PutWarmPool with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minSize:PutWarmPool', putWarmPool_minSize - Specifies the minimum number of instances to maintain in the warm pool. This helps you to ensure that there is always a certain number of warmed instances available to handle traffic spikes. Defaults to 0 if not specified.

$sel:maxGroupPreparedCapacity:PutWarmPool', putWarmPool_maxGroupPreparedCapacity - Specifies the maximum number of instances that are allowed to be in the warm pool or in any state except Terminated for the Auto Scaling group. This is an optional property. Specify it only if you do not want the warm pool size to be determined by the difference between the group's maximum capacity and its desired capacity.

If a value for MaxGroupPreparedCapacity is not specified, Amazon EC2 Auto Scaling launches and maintains the difference between the group's maximum capacity and its desired capacity. If you specify a value for MaxGroupPreparedCapacity, Amazon EC2 Auto Scaling uses the difference between the MaxGroupPreparedCapacity and the desired capacity instead.

The size of the warm pool is dynamic. Only when MaxGroupPreparedCapacity and MinSize are set to the same value does the warm pool have an absolute size.

If the desired capacity of the Auto Scaling group is higher than the MaxGroupPreparedCapacity, the capacity of the warm pool is 0, unless you specify a value for MinSize. To remove a value that you previously set, include the property but specify -1 for the value.

$sel:poolState:PutWarmPool', putWarmPool_poolState - Sets the instance state to transition to after the lifecycle actions are complete. Default is Stopped.

$sel:autoScalingGroupName:PutWarmPool', putWarmPool_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

putWarmPool_minSize :: Lens' PutWarmPool (Maybe Natural)

Specifies the minimum number of instances to maintain in the warm pool. This helps you to ensure that there is always a certain number of warmed instances available to handle traffic spikes. Defaults to 0 if not specified.

putWarmPool_maxGroupPreparedCapacity :: Lens' PutWarmPool (Maybe Int)

Specifies the maximum number of instances that are allowed to be in the warm pool or in any state except Terminated for the Auto Scaling group. This is an optional property. Specify it only if you do not want the warm pool size to be determined by the difference between the group's maximum capacity and its desired capacity.

If a value for MaxGroupPreparedCapacity is not specified, Amazon EC2 Auto Scaling launches and maintains the difference between the group's maximum capacity and its desired capacity. If you specify a value for MaxGroupPreparedCapacity, Amazon EC2 Auto Scaling uses the difference between the MaxGroupPreparedCapacity and the desired capacity instead.

The size of the warm pool is dynamic. Only when MaxGroupPreparedCapacity and MinSize are set to the same value does the warm pool have an absolute size.

If the desired capacity of the Auto Scaling group is higher than the MaxGroupPreparedCapacity, the capacity of the warm pool is 0, unless you specify a value for MinSize. To remove a value that you previously set, include the property but specify -1 for the value.

putWarmPool_poolState :: Lens' PutWarmPool (Maybe WarmPoolState)

Sets the instance state to transition to after the lifecycle actions are complete. Default is Stopped.

putWarmPool_autoScalingGroupName :: Lens' PutWarmPool Text

The name of the Auto Scaling group.

Destructuring the Response

data PutWarmPoolResponse

See: newPutWarmPoolResponse smart constructor.

Constructors

PutWarmPoolResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutWarmPoolResponse

Arguments

:: Int
$sel:httpStatus:PutWarmPoolResponse'

-> PutWarmPoolResponse

Create a value of PutWarmPoolResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutWarmPoolResponse', putWarmPoolResponse_httpStatus - The response's http status code.

Response Lenses

putWarmPoolResponse_httpStatus :: Lens' PutWarmPoolResponse Int

The response's http status code.