Network.AWS.AutoScaling.RecordLifecycleActionHeartbeat

Description

Records a heartbeat for the lifecycle action associated with the specified token or instance. This extends the timeout by the length of time defined using the PutLifecycleHook API call.

This step is a part of the procedure for adding a lifecycle hook to an Auto Scaling group:

(Optional) Create a Lambda function and a rule that allows CloudWatch Events to invoke your Lambda function when Amazon EC2 Auto Scaling launches or terminates instances.

(Optional) Create a notification target and an IAM role. The target can be either an Amazon SQS queue or an Amazon SNS topic. The role allows Amazon EC2 Auto Scaling to publish lifecycle notifications to the target.

Create the lifecycle hook. Specify whether the hook is used when the instances launch or terminate.

__If you need more time, record the lifecycle action heartbeat to keep the instance in a pending state.__

If you finish before the timeout period ends, complete the lifecycle action.

For more information, see Amazon EC2 Auto Scaling lifecycle hooks in the Amazon EC2 Auto Scaling User Guide.

Creating a Request

data RecordLifecycleActionHeartbeat

See: newRecordLifecycleActionHeartbeat smart constructor.

Constructors

RecordLifecycleActionHeartbeat'

Fields

instanceId :: Maybe Text

The ID of the instance.

lifecycleActionToken :: Maybe Text

A token that uniquely identifies a specific lifecycle action associated with an instance. Amazon EC2 Auto Scaling sends this token to the notification target that you specified when you created the lifecycle hook.

lifecycleHookName :: Text

The name of the lifecycle hook.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

newRecordLifecycleActionHeartbeat

Arguments

:: Text
$sel:lifecycleHookName:RecordLifecycleActionHeartbeat'

-> Text
$sel:autoScalingGroupName:RecordLifecycleActionHeartbeat'

-> RecordLifecycleActionHeartbeat

Create a value of RecordLifecycleActionHeartbeat with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:RecordLifecycleActionHeartbeat', recordLifecycleActionHeartbeat_instanceId - The ID of the instance.

$sel:lifecycleActionToken:RecordLifecycleActionHeartbeat', recordLifecycleActionHeartbeat_lifecycleActionToken - A token that uniquely identifies a specific lifecycle action associated with an instance. Amazon EC2 Auto Scaling sends this token to the notification target that you specified when you created the lifecycle hook.

$sel:lifecycleHookName:RecordLifecycleActionHeartbeat', recordLifecycleActionHeartbeat_lifecycleHookName - The name of the lifecycle hook.

$sel:autoScalingGroupName:RecordLifecycleActionHeartbeat', recordLifecycleActionHeartbeat_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

recordLifecycleActionHeartbeat_instanceId :: Lens' RecordLifecycleActionHeartbeat (Maybe Text)

The ID of the instance.

recordLifecycleActionHeartbeat_lifecycleActionToken :: Lens' RecordLifecycleActionHeartbeat (Maybe Text)

A token that uniquely identifies a specific lifecycle action associated with an instance. Amazon EC2 Auto Scaling sends this token to the notification target that you specified when you created the lifecycle hook.

recordLifecycleActionHeartbeat_lifecycleHookName :: Lens' RecordLifecycleActionHeartbeat Text

The name of the lifecycle hook.

recordLifecycleActionHeartbeat_autoScalingGroupName :: Lens' RecordLifecycleActionHeartbeat Text

The name of the Auto Scaling group.

Destructuring the Response

data RecordLifecycleActionHeartbeatResponse

See: newRecordLifecycleActionHeartbeatResponse smart constructor.

Constructors

RecordLifecycleActionHeartbeatResponse'

Fields

httpStatus :: Int

The response's http status code.

newRecordLifecycleActionHeartbeatResponse

Arguments

:: Int
$sel:httpStatus:RecordLifecycleActionHeartbeatResponse'

-> RecordLifecycleActionHeartbeatResponse

Create a value of RecordLifecycleActionHeartbeatResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RecordLifecycleActionHeartbeatResponse', recordLifecycleActionHeartbeatResponse_httpStatus - The response's http status code.

Response Lenses

recordLifecycleActionHeartbeatResponse_httpStatus :: Lens' RecordLifecycleActionHeartbeatResponse Int

The response's http status code.