Network.AWS.AutoScaling.SetInstanceHealth

Description

Sets the health status of the specified instance.

For more information, see Health checks for Auto Scaling instances in the Amazon EC2 Auto Scaling User Guide.

Creating a Request

data SetInstanceHealth

See: newSetInstanceHealth smart constructor.

Constructors

SetInstanceHealth'

Fields

shouldRespectGracePeriod :: Maybe Bool

If the Auto Scaling group of the specified instance has a HealthCheckGracePeriod specified for the group, by default, this call respects the grace period. Set this to False, to have the call not respect the grace period associated with the group.

For more information about the health check grace period, see CreateAutoScalingGroup in the Amazon EC2 Auto Scaling API Reference.

instanceId :: Text

The ID of the instance.

healthStatus :: Text

The health status of the instance. Set to Healthy to have the instance remain in service. Set to Unhealthy to have the instance be out of service. Amazon EC2 Auto Scaling terminates and replaces the unhealthy instance.

newSetInstanceHealth

Arguments

:: Text
$sel:instanceId:SetInstanceHealth'

-> Text
$sel:healthStatus:SetInstanceHealth'

-> SetInstanceHealth

Create a value of SetInstanceHealth with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shouldRespectGracePeriod:SetInstanceHealth', setInstanceHealth_shouldRespectGracePeriod - If the Auto Scaling group of the specified instance has a HealthCheckGracePeriod specified for the group, by default, this call respects the grace period. Set this to False, to have the call not respect the grace period associated with the group.

For more information about the health check grace period, see CreateAutoScalingGroup in the Amazon EC2 Auto Scaling API Reference.

$sel:instanceId:SetInstanceHealth', setInstanceHealth_instanceId - The ID of the instance.

$sel:healthStatus:SetInstanceHealth', setInstanceHealth_healthStatus - The health status of the instance. Set to Healthy to have the instance remain in service. Set to Unhealthy to have the instance be out of service. Amazon EC2 Auto Scaling terminates and replaces the unhealthy instance.

Request Lenses

setInstanceHealth_shouldRespectGracePeriod :: Lens' SetInstanceHealth (Maybe Bool)

If the Auto Scaling group of the specified instance has a HealthCheckGracePeriod specified for the group, by default, this call respects the grace period. Set this to False, to have the call not respect the grace period associated with the group.

For more information about the health check grace period, see CreateAutoScalingGroup in the Amazon EC2 Auto Scaling API Reference.

setInstanceHealth_instanceId :: Lens' SetInstanceHealth Text

The ID of the instance.

setInstanceHealth_healthStatus :: Lens' SetInstanceHealth Text

The health status of the instance. Set to Healthy to have the instance remain in service. Set to Unhealthy to have the instance be out of service. Amazon EC2 Auto Scaling terminates and replaces the unhealthy instance.

Destructuring the Response

data SetInstanceHealthResponse

See: newSetInstanceHealthResponse smart constructor.

Constructors

SetInstanceHealthResponse'

newSetInstanceHealthResponse :: SetInstanceHealthResponse

Create a value of SetInstanceHealthResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.