Network.AWS.AutoScaling.SetInstanceProtection

Description

Updates the instance protection settings of the specified instances. This operation cannot be called on instances in a warm pool.

For more information about preventing instances that are part of an Auto Scaling group from terminating on scale in, see Instance scale-in protection in the Amazon EC2 Auto Scaling User Guide.

If you exceed your maximum limit of instance IDs, which is 50 per Auto Scaling group, the call fails.

Creating a Request

data SetInstanceProtection

See: newSetInstanceProtection smart constructor.

Constructors

SetInstanceProtection'

Fields

instanceIds :: [Text]

One or more instance IDs. You can specify up to 50 instances.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

protectedFromScaleIn :: Bool

Indicates whether the instance is protected from termination by Amazon EC2 Auto Scaling when scaling in.

newSetInstanceProtection

Arguments

:: Text
$sel:autoScalingGroupName:SetInstanceProtection'

-> Bool
$sel:protectedFromScaleIn:SetInstanceProtection'

-> SetInstanceProtection

Create a value of SetInstanceProtection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceIds:SetInstanceProtection', setInstanceProtection_instanceIds - One or more instance IDs. You can specify up to 50 instances.

$sel:autoScalingGroupName:SetInstanceProtection', setInstanceProtection_autoScalingGroupName - The name of the Auto Scaling group.

$sel:protectedFromScaleIn:SetInstanceProtection', setInstanceProtection_protectedFromScaleIn - Indicates whether the instance is protected from termination by Amazon EC2 Auto Scaling when scaling in.

Request Lenses

setInstanceProtection_instanceIds :: Lens' SetInstanceProtection [Text]

One or more instance IDs. You can specify up to 50 instances.

setInstanceProtection_autoScalingGroupName :: Lens' SetInstanceProtection Text

The name of the Auto Scaling group.

setInstanceProtection_protectedFromScaleIn :: Lens' SetInstanceProtection Bool

Indicates whether the instance is protected from termination by Amazon EC2 Auto Scaling when scaling in.

Destructuring the Response

data SetInstanceProtectionResponse

See: newSetInstanceProtectionResponse smart constructor.

Constructors

SetInstanceProtectionResponse'

Fields

httpStatus :: Int

The response's http status code.

newSetInstanceProtectionResponse

Arguments

:: Int
$sel:httpStatus:SetInstanceProtectionResponse'

-> SetInstanceProtectionResponse

Create a value of SetInstanceProtectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SetInstanceProtectionResponse', setInstanceProtectionResponse_httpStatus - The response's http status code.

Response Lenses

setInstanceProtectionResponse_httpStatus :: Lens' SetInstanceProtectionResponse Int

The response's http status code.