Network.AWS.AutoScaling.SuspendProcesses

Description

Suspends the specified auto scaling processes, or all processes, for the specified Auto Scaling group.

If you suspend either the Launch or Terminate process types, it can prevent other process types from functioning properly. For more information, see Suspending and resuming scaling processes in the Amazon EC2 Auto Scaling User Guide.

To resume processes that have been suspended, call the ResumeProcesses API.

Creating a Request

data SuspendProcesses

See: newSuspendProcesses smart constructor.

Constructors

SuspendProcesses'

Fields

scalingProcesses :: Maybe [Text]

One or more of the following processes:

Launch

Terminate

AddToLoadBalancer

AlarmNotification

AZRebalance

HealthCheck

InstanceRefresh

ReplaceUnhealthy

ScheduledActions

If you omit this parameter, all processes are specified.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

newSuspendProcesses

Arguments

:: Text
$sel:autoScalingGroupName:SuspendProcesses'

-> SuspendProcesses

Create a value of SuspendProcesses with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingProcesses:SuspendProcesses', suspendProcesses_scalingProcesses - One or more of the following processes:

Launch

Terminate

AddToLoadBalancer

AlarmNotification

AZRebalance

HealthCheck

InstanceRefresh

ReplaceUnhealthy

ScheduledActions

If you omit this parameter, all processes are specified.

$sel:autoScalingGroupName:SuspendProcesses', suspendProcesses_autoScalingGroupName - The name of the Auto Scaling group.

Request Lenses

suspendProcesses_scalingProcesses :: Lens' SuspendProcesses (Maybe [Text])

One or more of the following processes:

Launch

Terminate

AddToLoadBalancer

AlarmNotification

AZRebalance

HealthCheck

InstanceRefresh

ReplaceUnhealthy

ScheduledActions

If you omit this parameter, all processes are specified.

suspendProcesses_autoScalingGroupName :: Lens' SuspendProcesses Text

The name of the Auto Scaling group.

Destructuring the Response

data SuspendProcessesResponse

See: newSuspendProcessesResponse smart constructor.

Constructors

SuspendProcessesResponse'

newSuspendProcessesResponse :: SuspendProcessesResponse

Create a value of SuspendProcessesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.