Network.AWS.AutoScaling.Types.Activity

Description

Documentation

data Activity

Describes scaling activity, which is a long-running process that represents a change to your Auto Scaling group, such as changing its size or replacing an instance.

See: newActivity smart constructor.

Constructors

Activity'

Fields

statusMessage :: Maybe Text

A friendly, more verbose description of the activity status.

autoScalingGroupARN :: Maybe Text

The Amazon Resource Name (ARN) of the Auto Scaling group.

endTime :: Maybe ISO8601

The end time of the activity.

autoScalingGroupState :: Maybe Text

The state of the Auto Scaling group, which is either InService or Deleted.

details :: Maybe Text

The details about the activity.

description :: Maybe Text

A friendly, more verbose description of the activity.

progress :: Maybe Int

A value between 0 and 100 that indicates the progress of the activity.

autoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

activityId :: Text

The ID of the activity.

cause :: Text

The reason the activity began.

startTime :: ISO8601

The start time of the activity.

statusCode :: ScalingActivityStatusCode

The current status of the activity.

newActivity

Arguments

:: Text
$sel:activityId:Activity'

-> Text
$sel:cause:Activity'

-> UTCTime
$sel:startTime:Activity'

-> ScalingActivityStatusCode
$sel:statusCode:Activity'

-> Activity

Create a value of Activity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:Activity', activity_statusMessage - A friendly, more verbose description of the activity status.

$sel:autoScalingGroupARN:Activity', activity_autoScalingGroupARN - The Amazon Resource Name (ARN) of the Auto Scaling group.

$sel:endTime:Activity', activity_endTime - The end time of the activity.

$sel:autoScalingGroupState:Activity', activity_autoScalingGroupState - The state of the Auto Scaling group, which is either InService or Deleted.

$sel:details:Activity', activity_details - The details about the activity.

$sel:description:Activity', activity_description - A friendly, more verbose description of the activity.

$sel:progress:Activity', activity_progress - A value between 0 and 100 that indicates the progress of the activity.

$sel:autoScalingGroupName:Activity', activity_autoScalingGroupName - The name of the Auto Scaling group.

$sel:activityId:Activity', activity_activityId - The ID of the activity.

$sel:cause:Activity', activity_cause - The reason the activity began.

$sel:startTime:Activity', activity_startTime - The start time of the activity.

$sel:statusCode:Activity', activity_statusCode - The current status of the activity.

activity_statusMessage :: Lens' Activity (Maybe Text)

A friendly, more verbose description of the activity status.

activity_autoScalingGroupARN :: Lens' Activity (Maybe Text)

The Amazon Resource Name (ARN) of the Auto Scaling group.

activity_endTime :: Lens' Activity (Maybe UTCTime)

The end time of the activity.

activity_autoScalingGroupState :: Lens' Activity (Maybe Text)

The state of the Auto Scaling group, which is either InService or Deleted.

activity_details :: Lens' Activity (Maybe Text)

The details about the activity.

activity_description :: Lens' Activity (Maybe Text)

A friendly, more verbose description of the activity.

activity_progress :: Lens' Activity (Maybe Int)

A value between 0 and 100 that indicates the progress of the activity.

activity_autoScalingGroupName :: Lens' Activity (Maybe Text)

The name of the Auto Scaling group.

activity_activityId :: Lens' Activity Text

The ID of the activity.

activity_cause :: Lens' Activity Text

The reason the activity began.

activity_startTime :: Lens' Activity UTCTime

The start time of the activity.

activity_statusCode :: Lens' Activity ScalingActivityStatusCode

The current status of the activity.