Network.AWS.AutoScaling.Types.AutoScalingGroup

Description

Documentation

data AutoScalingGroup

Describes an Auto Scaling group.

See: newAutoScalingGroup smart constructor.

Constructors

AutoScalingGroup'

Fields

status :: Maybe Text

The current state of the group when the DeleteAutoScalingGroup operation is in progress.

maxInstanceLifetime :: Maybe Int

The maximum amount of time, in seconds, that an instance can be in service.

Valid Range: Minimum value of 0.

placementGroup :: Maybe Text

The name of the placement group into which to launch your instances, if any.

suspendedProcesses :: Maybe [SuspendedProcess]

The suspended processes associated with the group.

predictedCapacity :: Maybe Int

The predicted capacity of the group when it has a predictive scaling policy.

autoScalingGroupARN :: Maybe Text

The Amazon Resource Name (ARN) of the Auto Scaling group.

launchTemplate :: Maybe LaunchTemplateSpecification

The launch template for the group.

launchConfigurationName :: Maybe Text

The name of the associated launch configuration.

instances :: Maybe [Instance]

The EC2 instances associated with the group.

context :: Maybe Text

Reserved.

mixedInstancesPolicy :: Maybe MixedInstancesPolicy

The mixed instances policy for the group.

tags :: Maybe [TagDescription]

The tags for the group.

loadBalancerNames :: Maybe [Text]

One or more load balancers associated with the group.

targetGroupARNs :: Maybe [Text]

The Amazon Resource Names (ARN) of the target groups for your load balancer.

vPCZoneIdentifier :: Maybe Text

One or more subnet IDs, if applicable, separated by commas.

warmPoolSize :: Maybe Int

The current size of the warm pool.

warmPoolConfiguration :: Maybe WarmPoolConfiguration

The warm pool for the group.

capacityRebalance :: Maybe Bool

Indicates whether Capacity Rebalancing is enabled.

serviceLinkedRoleARN :: Maybe Text

The Amazon Resource Name (ARN) of the service-linked role that the Auto Scaling group uses to call other Amazon Web Services on your behalf.

healthCheckGracePeriod :: Maybe Int

The amount of time, in seconds, that Amazon EC2 Auto Scaling waits before checking the health status of an EC2 instance that has come into service.

newInstancesProtectedFromScaleIn' :: Maybe Bool

Indicates whether newly launched instances are protected from termination by Amazon EC2 Auto Scaling when scaling in.

enabledMetrics :: Maybe [EnabledMetric]

The metrics enabled for the group.

terminationPolicies :: Maybe [Text]

The termination policies for the group.

autoScalingGroupName :: Text

The name of the Auto Scaling group.

minSize :: Int

The minimum size of the group.

maxSize :: Int

The maximum size of the group.

desiredCapacity :: Int

The desired size of the group.

defaultCooldown :: Int

The duration of the default cooldown period, in seconds.

availabilityZones :: [Text]

One or more Availability Zones for the group.

healthCheckType :: Text

The service to use for the health checks. The valid values are EC2 and ELB. If you configure an Auto Scaling group to use ELB health checks, it considers the instance unhealthy if it fails either the EC2 status checks or the load balancer health checks.

createdTime :: ISO8601

The date and time the group was created.

newAutoScalingGroup

Arguments

:: Text
$sel:autoScalingGroupName:AutoScalingGroup'

-> Int
$sel:minSize:AutoScalingGroup'

-> Int
$sel:maxSize:AutoScalingGroup'

-> Int
$sel:desiredCapacity:AutoScalingGroup'

-> Int
$sel:defaultCooldown:AutoScalingGroup'

-> Text
$sel:healthCheckType:AutoScalingGroup'

-> UTCTime
$sel:createdTime:AutoScalingGroup'

-> AutoScalingGroup

Create a value of AutoScalingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AutoScalingGroup', autoScalingGroup_status - The current state of the group when the DeleteAutoScalingGroup operation is in progress.

$sel:maxInstanceLifetime:AutoScalingGroup', autoScalingGroup_maxInstanceLifetime - The maximum amount of time, in seconds, that an instance can be in service.

Valid Range: Minimum value of 0.

$sel:placementGroup:AutoScalingGroup', autoScalingGroup_placementGroup - The name of the placement group into which to launch your instances, if any.

$sel:suspendedProcesses:AutoScalingGroup', autoScalingGroup_suspendedProcesses - The suspended processes associated with the group.

$sel:predictedCapacity:AutoScalingGroup', autoScalingGroup_predictedCapacity - The predicted capacity of the group when it has a predictive scaling policy.

$sel:autoScalingGroupARN:AutoScalingGroup', autoScalingGroup_autoScalingGroupARN - The Amazon Resource Name (ARN) of the Auto Scaling group.

$sel:launchTemplate:AutoScalingGroup', autoScalingGroup_launchTemplate - The launch template for the group.

$sel:launchConfigurationName:AutoScalingGroup', autoScalingGroup_launchConfigurationName - The name of the associated launch configuration.

$sel:instances:AutoScalingGroup', autoScalingGroup_instances - The EC2 instances associated with the group.

$sel:context:AutoScalingGroup', autoScalingGroup_context - Reserved.

$sel:mixedInstancesPolicy:AutoScalingGroup', autoScalingGroup_mixedInstancesPolicy - The mixed instances policy for the group.

$sel:tags:AutoScalingGroup', autoScalingGroup_tags - The tags for the group.

$sel:loadBalancerNames:AutoScalingGroup', autoScalingGroup_loadBalancerNames - One or more load balancers associated with the group.

$sel:targetGroupARNs:AutoScalingGroup', autoScalingGroup_targetGroupARNs - The Amazon Resource Names (ARN) of the target groups for your load balancer.

$sel:vPCZoneIdentifier:AutoScalingGroup', autoScalingGroup_vPCZoneIdentifier - One or more subnet IDs, if applicable, separated by commas.

$sel:warmPoolSize:AutoScalingGroup', autoScalingGroup_warmPoolSize - The current size of the warm pool.

$sel:warmPoolConfiguration:AutoScalingGroup', autoScalingGroup_warmPoolConfiguration - The warm pool for the group.

$sel:capacityRebalance:AutoScalingGroup', autoScalingGroup_capacityRebalance - Indicates whether Capacity Rebalancing is enabled.

$sel:serviceLinkedRoleARN:AutoScalingGroup', autoScalingGroup_serviceLinkedRoleARN - The Amazon Resource Name (ARN) of the service-linked role that the Auto Scaling group uses to call other Amazon Web Services on your behalf.

$sel:healthCheckGracePeriod:AutoScalingGroup', autoScalingGroup_healthCheckGracePeriod - The amount of time, in seconds, that Amazon EC2 Auto Scaling waits before checking the health status of an EC2 instance that has come into service.

$sel:newInstancesProtectedFromScaleIn':AutoScalingGroup', autoScalingGroup_newInstancesProtectedFromScaleIn - Indicates whether newly launched instances are protected from termination by Amazon EC2 Auto Scaling when scaling in.

$sel:enabledMetrics:AutoScalingGroup', autoScalingGroup_enabledMetrics - The metrics enabled for the group.

$sel:terminationPolicies:AutoScalingGroup', autoScalingGroup_terminationPolicies - The termination policies for the group.

$sel:autoScalingGroupName:AutoScalingGroup', autoScalingGroup_autoScalingGroupName - The name of the Auto Scaling group.

$sel:minSize:AutoScalingGroup', autoScalingGroup_minSize - The minimum size of the group.

$sel:maxSize:AutoScalingGroup', autoScalingGroup_maxSize - The maximum size of the group.

$sel:desiredCapacity:AutoScalingGroup', autoScalingGroup_desiredCapacity - The desired size of the group.

$sel:defaultCooldown:AutoScalingGroup', autoScalingGroup_defaultCooldown - The duration of the default cooldown period, in seconds.

$sel:availabilityZones:AutoScalingGroup', autoScalingGroup_availabilityZones - One or more Availability Zones for the group.

$sel:healthCheckType:AutoScalingGroup', autoScalingGroup_healthCheckType - The service to use for the health checks. The valid values are EC2 and ELB. If you configure an Auto Scaling group to use ELB health checks, it considers the instance unhealthy if it fails either the EC2 status checks or the load balancer health checks.

$sel:createdTime:AutoScalingGroup', autoScalingGroup_createdTime - The date and time the group was created.

autoScalingGroup_status :: Lens' AutoScalingGroup (Maybe Text)

The current state of the group when the DeleteAutoScalingGroup operation is in progress.

autoScalingGroup_maxInstanceLifetime :: Lens' AutoScalingGroup (Maybe Int)

The maximum amount of time, in seconds, that an instance can be in service.

Valid Range: Minimum value of 0.

autoScalingGroup_placementGroup :: Lens' AutoScalingGroup (Maybe Text)

The name of the placement group into which to launch your instances, if any.

autoScalingGroup_suspendedProcesses :: Lens' AutoScalingGroup (Maybe [SuspendedProcess])

The suspended processes associated with the group.

autoScalingGroup_predictedCapacity :: Lens' AutoScalingGroup (Maybe Int)

The predicted capacity of the group when it has a predictive scaling policy.

autoScalingGroup_autoScalingGroupARN :: Lens' AutoScalingGroup (Maybe Text)

The Amazon Resource Name (ARN) of the Auto Scaling group.

autoScalingGroup_launchTemplate :: Lens' AutoScalingGroup (Maybe LaunchTemplateSpecification)

The launch template for the group.

autoScalingGroup_launchConfigurationName :: Lens' AutoScalingGroup (Maybe Text)

The name of the associated launch configuration.

autoScalingGroup_instances :: Lens' AutoScalingGroup (Maybe [Instance])

The EC2 instances associated with the group.

autoScalingGroup_context :: Lens' AutoScalingGroup (Maybe Text)

Reserved.

autoScalingGroup_mixedInstancesPolicy :: Lens' AutoScalingGroup (Maybe MixedInstancesPolicy)

The mixed instances policy for the group.

autoScalingGroup_tags :: Lens' AutoScalingGroup (Maybe [TagDescription])

The tags for the group.

autoScalingGroup_loadBalancerNames :: Lens' AutoScalingGroup (Maybe [Text])

One or more load balancers associated with the group.

autoScalingGroup_targetGroupARNs :: Lens' AutoScalingGroup (Maybe [Text])

The Amazon Resource Names (ARN) of the target groups for your load balancer.

autoScalingGroup_vPCZoneIdentifier :: Lens' AutoScalingGroup (Maybe Text)

One or more subnet IDs, if applicable, separated by commas.

autoScalingGroup_warmPoolSize :: Lens' AutoScalingGroup (Maybe Int)

The current size of the warm pool.

autoScalingGroup_warmPoolConfiguration :: Lens' AutoScalingGroup (Maybe WarmPoolConfiguration)

The warm pool for the group.

autoScalingGroup_capacityRebalance :: Lens' AutoScalingGroup (Maybe Bool)

Indicates whether Capacity Rebalancing is enabled.

autoScalingGroup_serviceLinkedRoleARN :: Lens' AutoScalingGroup (Maybe Text)

The Amazon Resource Name (ARN) of the service-linked role that the Auto Scaling group uses to call other Amazon Web Services on your behalf.

autoScalingGroup_healthCheckGracePeriod :: Lens' AutoScalingGroup (Maybe Int)

The amount of time, in seconds, that Amazon EC2 Auto Scaling waits before checking the health status of an EC2 instance that has come into service.

autoScalingGroup_newInstancesProtectedFromScaleIn :: Lens' AutoScalingGroup (Maybe Bool)

Indicates whether newly launched instances are protected from termination by Amazon EC2 Auto Scaling when scaling in.

autoScalingGroup_enabledMetrics :: Lens' AutoScalingGroup (Maybe [EnabledMetric])

The metrics enabled for the group.

autoScalingGroup_terminationPolicies :: Lens' AutoScalingGroup (Maybe [Text])

The termination policies for the group.

autoScalingGroup_autoScalingGroupName :: Lens' AutoScalingGroup Text

The name of the Auto Scaling group.

autoScalingGroup_minSize :: Lens' AutoScalingGroup Int

The minimum size of the group.

autoScalingGroup_maxSize :: Lens' AutoScalingGroup Int

The maximum size of the group.

autoScalingGroup_desiredCapacity :: Lens' AutoScalingGroup Int

The desired size of the group.

autoScalingGroup_defaultCooldown :: Lens' AutoScalingGroup Int

The duration of the default cooldown period, in seconds.

autoScalingGroup_availabilityZones :: Lens' AutoScalingGroup [Text]

One or more Availability Zones for the group.

autoScalingGroup_healthCheckType :: Lens' AutoScalingGroup Text

The service to use for the health checks. The valid values are EC2 and ELB. If you configure an Auto Scaling group to use ELB health checks, it considers the instance unhealthy if it fails either the EC2 status checks or the load balancer health checks.

autoScalingGroup_createdTime :: Lens' AutoScalingGroup UTCTime

The date and time the group was created.