Network.AWS.AutoScaling.Types.BlockDeviceMapping

Description

Documentation

data BlockDeviceMapping

Describes a block device mapping.

See: newBlockDeviceMapping smart constructor.

Constructors

BlockDeviceMapping'

Fields

virtualName :: Maybe Text

The name of the virtual device (for example, ephemeral0).

You can specify either VirtualName or Ebs, but not both.

noDevice :: Maybe Bool

Setting this value to true suppresses the specified device included in the block device mapping of the AMI.

If NoDevice is true for the root device, instances might fail the EC2 health check. In that case, Amazon EC2 Auto Scaling launches replacement instances.

If you specify NoDevice, you cannot specify Ebs.

ebs :: Maybe Ebs

Parameters used to automatically set up EBS volumes when an instance is launched.

You can specify either VirtualName or Ebs, but not both.

deviceName :: Text

The device name exposed to the EC2 instance (for example, /dev/sdh or xvdh). For more information, see Device Naming on Linux Instances in the Amazon EC2 User Guide for Linux Instances.

newBlockDeviceMapping

Arguments

:: Text
$sel:deviceName:BlockDeviceMapping'

-> BlockDeviceMapping

Create a value of BlockDeviceMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:virtualName:BlockDeviceMapping', blockDeviceMapping_virtualName - The name of the virtual device (for example, ephemeral0).

You can specify either VirtualName or Ebs, but not both.

$sel:noDevice:BlockDeviceMapping', blockDeviceMapping_noDevice - Setting this value to true suppresses the specified device included in the block device mapping of the AMI.

If NoDevice is true for the root device, instances might fail the EC2 health check. In that case, Amazon EC2 Auto Scaling launches replacement instances.

If you specify NoDevice, you cannot specify Ebs.

$sel:ebs:BlockDeviceMapping', blockDeviceMapping_ebs - Parameters used to automatically set up EBS volumes when an instance is launched.

You can specify either VirtualName or Ebs, but not both.

$sel:deviceName:BlockDeviceMapping', blockDeviceMapping_deviceName - The device name exposed to the EC2 instance (for example, /dev/sdh or xvdh). For more information, see Device Naming on Linux Instances in the Amazon EC2 User Guide for Linux Instances.

blockDeviceMapping_virtualName :: Lens' BlockDeviceMapping (Maybe Text)

The name of the virtual device (for example, ephemeral0).

You can specify either VirtualName or Ebs, but not both.

blockDeviceMapping_noDevice :: Lens' BlockDeviceMapping (Maybe Bool)

Setting this value to true suppresses the specified device included in the block device mapping of the AMI.

If NoDevice is true for the root device, instances might fail the EC2 health check. In that case, Amazon EC2 Auto Scaling launches replacement instances.

If you specify NoDevice, you cannot specify Ebs.

blockDeviceMapping_ebs :: Lens' BlockDeviceMapping (Maybe Ebs)

Parameters used to automatically set up EBS volumes when an instance is launched.

You can specify either VirtualName or Ebs, but not both.

blockDeviceMapping_deviceName :: Lens' BlockDeviceMapping Text

The device name exposed to the EC2 instance (for example, /dev/sdh or xvdh). For more information, see Device Naming on Linux Instances in the Amazon EC2 User Guide for Linux Instances.