Network.AWS.AutoScaling.Types.DesiredConfiguration

Description

Documentation

data DesiredConfiguration

Describes the desired configuration for an instance refresh.

If you specify a desired configuration, you must specify either a LaunchTemplate or a MixedInstancesPolicy.

See: newDesiredConfiguration smart constructor.

Constructors

DesiredConfiguration'

Fields

launchTemplate :: Maybe LaunchTemplateSpecification

mixedInstancesPolicy :: Maybe MixedInstancesPolicy

newDesiredConfiguration :: DesiredConfiguration

Create a value of DesiredConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchTemplate:DesiredConfiguration', desiredConfiguration_launchTemplate - Undocumented member.

$sel:mixedInstancesPolicy:DesiredConfiguration', desiredConfiguration_mixedInstancesPolicy - Undocumented member.

desiredConfiguration_launchTemplate :: Lens' DesiredConfiguration (Maybe LaunchTemplateSpecification)

Undocumented member.

desiredConfiguration_mixedInstancesPolicy :: Lens' DesiredConfiguration (Maybe MixedInstancesPolicy)

Undocumented member.