Network.AWS.AutoScaling.Types.Ebs

Description

Documentation

data Ebs

Describes information used to set up an Amazon EBS volume specified in a block device mapping.

See: newEbs smart constructor.

Constructors

Ebs'

Fields

deleteOnTermination :: Maybe Bool

Indicates whether the volume is deleted on instance termination. For Amazon EC2 Auto Scaling, the default value is true.

throughput :: Maybe Natural

The throughput (MiBps) to provision for a gp3 volume.

volumeSize :: Maybe Natural

The volume size, in GiBs. The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

You must specify either a SnapshotId or a VolumeSize. If you specify both SnapshotId and VolumeSize, the volume size must be equal or greater than the size of the snapshot.

iops :: Maybe Natural

The number of input/output (I/O) operations per second (IOPS) to provision for the volume. For gp3 and io1 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

For io1 volumes, we guarantee 64,000 IOPS only for Instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS.

Iops is supported when the volume type is gp3 or io1 and required only when the volume type is io1. (Not used with standard, gp2, st1, or sc1 volumes.)

encrypted :: Maybe Bool

Specifies whether the volume should be encrypted. Encrypted EBS volumes can only be attached to instances that support Amazon EBS encryption. For more information, see Supported instance types. If your AMI uses encrypted volumes, you can also only launch it on supported instance types.

If you are creating a volume from a snapshot, you cannot create an unencrypted volume from an encrypted snapshot. Also, you cannot specify a KMS key ID when using a launch configuration.

If you enable encryption by default, the EBS volumes that you create are always encrypted, either using the Amazon Web Services managed KMS key or a customer-managed KMS key, regardless of whether the snapshot was encrypted.

For more information, see Using Amazon Web Services KMS keys to encrypt Amazon EBS volumes in the Amazon EC2 Auto Scaling User Guide.

volumeType :: Maybe Text

The volume type. For more information, see Amazon EBS volume types in the Amazon EC2 User Guide for Linux Instances.

Valid Values: standard | io1 | gp2 | st1 | sc1 | gp3

snapshotId :: Maybe Text

The snapshot ID of the volume to use.

You must specify either a VolumeSize or a SnapshotId.

newEbs :: Ebs

Create a value of Ebs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deleteOnTermination:Ebs', ebs_deleteOnTermination - Indicates whether the volume is deleted on instance termination. For Amazon EC2 Auto Scaling, the default value is true.

$sel:throughput:Ebs', ebs_throughput - The throughput (MiBps) to provision for a gp3 volume.

$sel:volumeSize:Ebs', ebs_volumeSize - The volume size, in GiBs. The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

You must specify either a SnapshotId or a VolumeSize. If you specify both SnapshotId and VolumeSize, the volume size must be equal or greater than the size of the snapshot.

$sel:iops:Ebs', ebs_iops - The number of input/output (I/O) operations per second (IOPS) to provision for the volume. For gp3 and io1 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

For io1 volumes, we guarantee 64,000 IOPS only for Instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS.

Iops is supported when the volume type is gp3 or io1 and required only when the volume type is io1. (Not used with standard, gp2, st1, or sc1 volumes.)

$sel:encrypted:Ebs', ebs_encrypted - Specifies whether the volume should be encrypted. Encrypted EBS volumes can only be attached to instances that support Amazon EBS encryption. For more information, see Supported instance types. If your AMI uses encrypted volumes, you can also only launch it on supported instance types.

If you are creating a volume from a snapshot, you cannot create an unencrypted volume from an encrypted snapshot. Also, you cannot specify a KMS key ID when using a launch configuration.

If you enable encryption by default, the EBS volumes that you create are always encrypted, either using the Amazon Web Services managed KMS key or a customer-managed KMS key, regardless of whether the snapshot was encrypted.

For more information, see Using Amazon Web Services KMS keys to encrypt Amazon EBS volumes in the Amazon EC2 Auto Scaling User Guide.

$sel:volumeType:Ebs', ebs_volumeType - The volume type. For more information, see Amazon EBS volume types in the Amazon EC2 User Guide for Linux Instances.

Valid Values: standard | io1 | gp2 | st1 | sc1 | gp3

$sel:snapshotId:Ebs', ebs_snapshotId - The snapshot ID of the volume to use.

You must specify either a VolumeSize or a SnapshotId.

ebs_deleteOnTermination :: Lens' Ebs (Maybe Bool)

Indicates whether the volume is deleted on instance termination. For Amazon EC2 Auto Scaling, the default value is true.

ebs_throughput :: Lens' Ebs (Maybe Natural)

The throughput (MiBps) to provision for a gp3 volume.

ebs_volumeSize :: Lens' Ebs (Maybe Natural)

The volume size, in GiBs. The following are the supported volumes sizes for each volume type:

gp2 and gp3: 1-16,384

io1: 4-16,384

st1 and sc1: 125-16,384

standard: 1-1,024

You must specify either a SnapshotId or a VolumeSize. If you specify both SnapshotId and VolumeSize, the volume size must be equal or greater than the size of the snapshot.

ebs_iops :: Lens' Ebs (Maybe Natural)

The number of input/output (I/O) operations per second (IOPS) to provision for the volume. For gp3 and io1 volumes, this represents the number of IOPS that are provisioned for the volume. For gp2 volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting.

The following are the supported values for each volume type:

gp3: 3,000-16,000 IOPS

io1: 100-64,000 IOPS

For io1 volumes, we guarantee 64,000 IOPS only for Instances built on the Nitro System. Other instance families guarantee performance up to 32,000 IOPS.

Iops is supported when the volume type is gp3 or io1 and required only when the volume type is io1. (Not used with standard, gp2, st1, or sc1 volumes.)

ebs_encrypted :: Lens' Ebs (Maybe Bool)

Specifies whether the volume should be encrypted. Encrypted EBS volumes can only be attached to instances that support Amazon EBS encryption. For more information, see Supported instance types. If your AMI uses encrypted volumes, you can also only launch it on supported instance types.

If you are creating a volume from a snapshot, you cannot create an unencrypted volume from an encrypted snapshot. Also, you cannot specify a KMS key ID when using a launch configuration.

If you enable encryption by default, the EBS volumes that you create are always encrypted, either using the Amazon Web Services managed KMS key or a customer-managed KMS key, regardless of whether the snapshot was encrypted.

For more information, see Using Amazon Web Services KMS keys to encrypt Amazon EBS volumes in the Amazon EC2 Auto Scaling User Guide.

ebs_volumeType :: Lens' Ebs (Maybe Text)

The volume type. For more information, see Amazon EBS volume types in the Amazon EC2 User Guide for Linux Instances.

Valid Values: standard | io1 | gp2 | st1 | sc1 | gp3

ebs_snapshotId :: Lens' Ebs (Maybe Text)

The snapshot ID of the volume to use.

You must specify either a VolumeSize or a SnapshotId.