Network.AWS.AutoScaling.Types.FailedScheduledUpdateGroupActionRequest

Description

Documentation

data FailedScheduledUpdateGroupActionRequest

Describes a scheduled action that could not be created, updated, or deleted.

See: newFailedScheduledUpdateGroupActionRequest smart constructor.

Constructors

FailedScheduledUpdateGroupActionRequest'

Fields

errorMessage :: Maybe Text

The error message accompanying the error code.

errorCode :: Maybe Text

The error code.

scheduledActionName :: Text

The name of the scheduled action.

newFailedScheduledUpdateGroupActionRequest

Arguments

:: Text
$sel:scheduledActionName:FailedScheduledUpdateGroupActionRequest'

-> FailedScheduledUpdateGroupActionRequest

Create a value of FailedScheduledUpdateGroupActionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorMessage:FailedScheduledUpdateGroupActionRequest', failedScheduledUpdateGroupActionRequest_errorMessage - The error message accompanying the error code.

$sel:errorCode:FailedScheduledUpdateGroupActionRequest', failedScheduledUpdateGroupActionRequest_errorCode - The error code.

$sel:scheduledActionName:FailedScheduledUpdateGroupActionRequest', failedScheduledUpdateGroupActionRequest_scheduledActionName - The name of the scheduled action.

failedScheduledUpdateGroupActionRequest_errorMessage :: Lens' FailedScheduledUpdateGroupActionRequest (Maybe Text)

The error message accompanying the error code.

failedScheduledUpdateGroupActionRequest_errorCode :: Lens' FailedScheduledUpdateGroupActionRequest (Maybe Text)

The error code.

failedScheduledUpdateGroupActionRequest_scheduledActionName :: Lens' FailedScheduledUpdateGroupActionRequest Text

The name of the scheduled action.