Network.AWS.AutoScaling.Types.Instance

Description

Documentation

data Instance

Describes an EC2 instance.

See: newInstance smart constructor.

Constructors

Instance'

Fields

instanceType :: Maybe Text

The instance type of the EC2 instance.

launchTemplate :: Maybe LaunchTemplateSpecification

The launch template for the instance.

launchConfigurationName :: Maybe Text

The launch configuration associated with the instance.

weightedCapacity :: Maybe Text

The number of capacity units contributed by the instance based on its instance type.

Valid Range: Minimum value of 1. Maximum value of 999.

instanceId :: Text

The ID of the instance.

availabilityZone :: Text

The Availability Zone in which the instance is running.

lifecycleState :: LifecycleState

A description of the current lifecycle state. The Quarantined state is not used. For information about lifecycle states, see Instance lifecycle in the Amazon EC2 Auto Scaling User Guide.

healthStatus :: Text

The last reported health status of the instance. "Healthy" means that the instance is healthy and should remain in service. "Unhealthy" means that the instance is unhealthy and that Amazon EC2 Auto Scaling should terminate and replace it.

protectedFromScaleIn :: Bool

Indicates whether the instance is protected from termination by Amazon EC2 Auto Scaling when scaling in.

newInstance

Arguments

:: Text
$sel:instanceId:Instance'

-> Text
$sel:availabilityZone:Instance'

-> LifecycleState
$sel:lifecycleState:Instance'

-> Text
$sel:healthStatus:Instance'

-> Bool
$sel:protectedFromScaleIn:Instance'

-> Instance

Create a value of Instance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceType:Instance', instance_instanceType - The instance type of the EC2 instance.

$sel:launchTemplate:Instance', instance_launchTemplate - The launch template for the instance.

$sel:launchConfigurationName:Instance', instance_launchConfigurationName - The launch configuration associated with the instance.

$sel:weightedCapacity:Instance', instance_weightedCapacity - The number of capacity units contributed by the instance based on its instance type.

Valid Range: Minimum value of 1. Maximum value of 999.

$sel:instanceId:Instance', instance_instanceId - The ID of the instance.

$sel:availabilityZone:Instance', instance_availabilityZone - The Availability Zone in which the instance is running.

$sel:lifecycleState:Instance', instance_lifecycleState - A description of the current lifecycle state. The Quarantined state is not used. For information about lifecycle states, see Instance lifecycle in the Amazon EC2 Auto Scaling User Guide.

$sel:healthStatus:Instance', instance_healthStatus - The last reported health status of the instance. "Healthy" means that the instance is healthy and should remain in service. "Unhealthy" means that the instance is unhealthy and that Amazon EC2 Auto Scaling should terminate and replace it.

$sel:protectedFromScaleIn:Instance', instance_protectedFromScaleIn - Indicates whether the instance is protected from termination by Amazon EC2 Auto Scaling when scaling in.

instance_instanceType :: Lens' Instance (Maybe Text)

The instance type of the EC2 instance.

instance_launchTemplate :: Lens' Instance (Maybe LaunchTemplateSpecification)

The launch template for the instance.

instance_launchConfigurationName :: Lens' Instance (Maybe Text)

The launch configuration associated with the instance.

instance_weightedCapacity :: Lens' Instance (Maybe Text)

The number of capacity units contributed by the instance based on its instance type.

Valid Range: Minimum value of 1. Maximum value of 999.

instance_instanceId :: Lens' Instance Text

The ID of the instance.

instance_availabilityZone :: Lens' Instance Text

The Availability Zone in which the instance is running.

instance_lifecycleState :: Lens' Instance LifecycleState

A description of the current lifecycle state. The Quarantined state is not used. For information about lifecycle states, see Instance lifecycle in the Amazon EC2 Auto Scaling User Guide.

instance_healthStatus :: Lens' Instance Text

The last reported health status of the instance. "Healthy" means that the instance is healthy and should remain in service. "Unhealthy" means that the instance is unhealthy and that Amazon EC2 Auto Scaling should terminate and replace it.

instance_protectedFromScaleIn :: Lens' Instance Bool

Indicates whether the instance is protected from termination by Amazon EC2 Auto Scaling when scaling in.