Network.AWS.AutoScaling.Types.InstanceRefresh

Description

Documentation

data InstanceRefresh

Describes an instance refresh for an Auto Scaling group.

See: newInstanceRefresh smart constructor.

Constructors

InstanceRefresh'

Fields

status :: Maybe InstanceRefreshStatus

The current status for the instance refresh operation:

Pending - The request was created, but the operation has not started.

InProgress - The operation is in progress.

Successful - The operation completed successfully.

Failed - The operation failed to complete. You can troubleshoot using the status reason and the scaling activities.

Cancelling - An ongoing operation is being cancelled. Cancellation does not roll back any replacements that have already been completed, but it prevents new replacements from being started.

Cancelled - The operation is cancelled.

instanceRefreshId :: Maybe Text

The instance refresh ID.

percentageComplete :: Maybe Natural

The percentage of the instance refresh that is complete. For each instance replacement, Amazon EC2 Auto Scaling tracks the instance's health status and warm-up time. When the instance's health status changes to healthy and the specified warm-up time passes, the instance is considered updated and is added to the percentage complete.

startTime :: Maybe ISO8601

The date and time at which the instance refresh began.

endTime :: Maybe ISO8601

The date and time at which the instance refresh ended.

progressDetails :: Maybe InstanceRefreshProgressDetails

Additional progress details for an Auto Scaling group that has a warm pool.

desiredConfiguration :: Maybe DesiredConfiguration

Describes the specific update you want to deploy.

preferences :: Maybe RefreshPreferences

statusReason :: Maybe Text

Provides more details about the current status of the instance refresh.

instancesToUpdate :: Maybe Natural

The number of instances remaining to update before the instance refresh is complete.

autoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

newInstanceRefresh :: InstanceRefresh

Create a value of InstanceRefresh with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:InstanceRefresh', instanceRefresh_status - The current status for the instance refresh operation:

Pending - The request was created, but the operation has not started.

InProgress - The operation is in progress.

Successful - The operation completed successfully.

Failed - The operation failed to complete. You can troubleshoot using the status reason and the scaling activities.

Cancelling - An ongoing operation is being cancelled. Cancellation does not roll back any replacements that have already been completed, but it prevents new replacements from being started.

Cancelled - The operation is cancelled.

$sel:instanceRefreshId:InstanceRefresh', instanceRefresh_instanceRefreshId - The instance refresh ID.

$sel:percentageComplete:InstanceRefresh', instanceRefresh_percentageComplete - The percentage of the instance refresh that is complete. For each instance replacement, Amazon EC2 Auto Scaling tracks the instance's health status and warm-up time. When the instance's health status changes to healthy and the specified warm-up time passes, the instance is considered updated and is added to the percentage complete.

$sel:startTime:InstanceRefresh', instanceRefresh_startTime - The date and time at which the instance refresh began.

$sel:endTime:InstanceRefresh', instanceRefresh_endTime - The date and time at which the instance refresh ended.

$sel:progressDetails:InstanceRefresh', instanceRefresh_progressDetails - Additional progress details for an Auto Scaling group that has a warm pool.

$sel:desiredConfiguration:InstanceRefresh', instanceRefresh_desiredConfiguration - Describes the specific update you want to deploy.

$sel:preferences:InstanceRefresh', instanceRefresh_preferences - Undocumented member.

$sel:statusReason:InstanceRefresh', instanceRefresh_statusReason - Provides more details about the current status of the instance refresh.

$sel:instancesToUpdate:InstanceRefresh', instanceRefresh_instancesToUpdate - The number of instances remaining to update before the instance refresh is complete.

$sel:autoScalingGroupName:InstanceRefresh', instanceRefresh_autoScalingGroupName - The name of the Auto Scaling group.

instanceRefresh_status :: Lens' InstanceRefresh (Maybe InstanceRefreshStatus)

The current status for the instance refresh operation:

Pending - The request was created, but the operation has not started.

InProgress - The operation is in progress.

Successful - The operation completed successfully.

Failed - The operation failed to complete. You can troubleshoot using the status reason and the scaling activities.

Cancelling - An ongoing operation is being cancelled. Cancellation does not roll back any replacements that have already been completed, but it prevents new replacements from being started.

Cancelled - The operation is cancelled.

instanceRefresh_instanceRefreshId :: Lens' InstanceRefresh (Maybe Text)

The instance refresh ID.

instanceRefresh_percentageComplete :: Lens' InstanceRefresh (Maybe Natural)

The percentage of the instance refresh that is complete. For each instance replacement, Amazon EC2 Auto Scaling tracks the instance's health status and warm-up time. When the instance's health status changes to healthy and the specified warm-up time passes, the instance is considered updated and is added to the percentage complete.

instanceRefresh_startTime :: Lens' InstanceRefresh (Maybe UTCTime)

The date and time at which the instance refresh began.

instanceRefresh_endTime :: Lens' InstanceRefresh (Maybe UTCTime)

The date and time at which the instance refresh ended.

instanceRefresh_progressDetails :: Lens' InstanceRefresh (Maybe InstanceRefreshProgressDetails)

Additional progress details for an Auto Scaling group that has a warm pool.

instanceRefresh_desiredConfiguration :: Lens' InstanceRefresh (Maybe DesiredConfiguration)

Describes the specific update you want to deploy.

instanceRefresh_preferences :: Lens' InstanceRefresh (Maybe RefreshPreferences)

Undocumented member.

instanceRefresh_statusReason :: Lens' InstanceRefresh (Maybe Text)

Provides more details about the current status of the instance refresh.

instanceRefresh_instancesToUpdate :: Lens' InstanceRefresh (Maybe Natural)

The number of instances remaining to update before the instance refresh is complete.

instanceRefresh_autoScalingGroupName :: Lens' InstanceRefresh (Maybe Text)

The name of the Auto Scaling group.