Network.AWS.AutoScaling.Types.LaunchConfiguration

Description

Documentation

data LaunchConfiguration

Describes a launch configuration.

See: newLaunchConfiguration smart constructor.

Constructors

LaunchConfiguration'

Fields

ebsOptimized :: Maybe Bool

Specifies whether the launch configuration is optimized for EBS I/O (true) or not (false). For more information, see Amazon EBS-Optimized Instances in the Amazon EC2 User Guide for Linux Instances.

userData :: Maybe Text

The user data to make available to the launched EC2 instances. For more information, see Instance metadata and user data (Linux) and Instance metadata and user data (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text. User data is limited to 16 KB.

ramdiskId :: Maybe Text

The ID of the RAM disk associated with the AMI.

classicLinkVPCSecurityGroups :: Maybe [Text]

The IDs of one or more security groups for the VPC specified in ClassicLinkVPCId.

For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

spotPrice :: Maybe Text

The maximum hourly price to be paid for any Spot Instance launched to fulfill the request. Spot Instances are launched when the price you specify exceeds the current Spot price. For more information, see Requesting Spot Instances in the Amazon EC2 Auto Scaling User Guide.

associatePublicIpAddress :: Maybe Bool

For Auto Scaling groups that are running in a VPC, specifies whether to assign a public IP address to the group's instances. For more information, see Launching Auto Scaling instances in a VPC in the Amazon EC2 Auto Scaling User Guide.

securityGroups :: Maybe [Text]

A list that contains the security groups to assign to the instances in the Auto Scaling group. For more information, see Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide.

iamInstanceProfile :: Maybe Text

The name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role. For more information, see IAM role for applications that run on Amazon EC2 instances in the Amazon EC2 Auto Scaling User Guide.

classicLinkVPCId :: Maybe Text

The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

blockDeviceMappings :: Maybe [BlockDeviceMapping]

A block device mapping, which specifies the block devices for the instance. For more information, see Block Device Mapping in the Amazon EC2 User Guide for Linux Instances.

kernelId :: Maybe Text

The ID of the kernel associated with the AMI.

placementTenancy :: Maybe Text

The tenancy of the instance, either default or dedicated. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

For more information, see Configuring instance tenancy with Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

launchConfigurationARN :: Maybe Text

The Amazon Resource Name (ARN) of the launch configuration.

keyName :: Maybe Text

The name of the key pair.

For more information, see Amazon EC2 Key Pairs in the Amazon EC2 User Guide for Linux Instances.

instanceMonitoring :: Maybe InstanceMonitoring

Controls whether instances in this group are launched with detailed (true) or basic (false) monitoring.

For more information, see Configure Monitoring for Auto Scaling Instances in the Amazon EC2 Auto Scaling User Guide.

metadataOptions :: Maybe InstanceMetadataOptions

The metadata options for the instances. For more information, see Configuring the Instance Metadata Options in the Amazon EC2 Auto Scaling User Guide.

launchConfigurationName :: Text

The name of the launch configuration.

imageId :: Text

The ID of the Amazon Machine Image (AMI) to use to launch your EC2 instances. For more information, see Finding an AMI in the Amazon EC2 User Guide for Linux Instances.

instanceType :: Text

The instance type for the instances.

For information about available instance types, see Available Instance Types in the Amazon EC2 User Guide for Linux Instances.

createdTime :: ISO8601

The creation date and time for the launch configuration.

newLaunchConfiguration

Arguments

:: Text
$sel:launchConfigurationName:LaunchConfiguration'

-> Text
$sel:imageId:LaunchConfiguration'

-> Text
$sel:instanceType:LaunchConfiguration'

-> UTCTime
$sel:createdTime:LaunchConfiguration'

-> LaunchConfiguration

Create a value of LaunchConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ebsOptimized:LaunchConfiguration', launchConfiguration_ebsOptimized - Specifies whether the launch configuration is optimized for EBS I/O (true) or not (false). For more information, see Amazon EBS-Optimized Instances in the Amazon EC2 User Guide for Linux Instances.

$sel:userData:LaunchConfiguration', launchConfiguration_userData - The user data to make available to the launched EC2 instances. For more information, see Instance metadata and user data (Linux) and Instance metadata and user data (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text. User data is limited to 16 KB.

$sel:ramdiskId:LaunchConfiguration', launchConfiguration_ramdiskId - The ID of the RAM disk associated with the AMI.

$sel:classicLinkVPCSecurityGroups:LaunchConfiguration', launchConfiguration_classicLinkVPCSecurityGroups - The IDs of one or more security groups for the VPC specified in ClassicLinkVPCId.

For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

$sel:spotPrice:LaunchConfiguration', launchConfiguration_spotPrice - The maximum hourly price to be paid for any Spot Instance launched to fulfill the request. Spot Instances are launched when the price you specify exceeds the current Spot price. For more information, see Requesting Spot Instances in the Amazon EC2 Auto Scaling User Guide.

$sel:associatePublicIpAddress:LaunchConfiguration', launchConfiguration_associatePublicIpAddress - For Auto Scaling groups that are running in a VPC, specifies whether to assign a public IP address to the group's instances. For more information, see Launching Auto Scaling instances in a VPC in the Amazon EC2 Auto Scaling User Guide.

$sel:securityGroups:LaunchConfiguration', launchConfiguration_securityGroups - A list that contains the security groups to assign to the instances in the Auto Scaling group. For more information, see Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide.

$sel:iamInstanceProfile:LaunchConfiguration', launchConfiguration_iamInstanceProfile - The name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role. For more information, see IAM role for applications that run on Amazon EC2 instances in the Amazon EC2 Auto Scaling User Guide.

$sel:classicLinkVPCId:LaunchConfiguration', launchConfiguration_classicLinkVPCId - The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

$sel:blockDeviceMappings:LaunchConfiguration', launchConfiguration_blockDeviceMappings - A block device mapping, which specifies the block devices for the instance. For more information, see Block Device Mapping in the Amazon EC2 User Guide for Linux Instances.

$sel:kernelId:LaunchConfiguration', launchConfiguration_kernelId - The ID of the kernel associated with the AMI.

$sel:placementTenancy:LaunchConfiguration', launchConfiguration_placementTenancy - The tenancy of the instance, either default or dedicated. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

For more information, see Configuring instance tenancy with Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

$sel:launchConfigurationARN:LaunchConfiguration', launchConfiguration_launchConfigurationARN - The Amazon Resource Name (ARN) of the launch configuration.

$sel:keyName:LaunchConfiguration', launchConfiguration_keyName - The name of the key pair.

For more information, see Amazon EC2 Key Pairs in the Amazon EC2 User Guide for Linux Instances.

$sel:instanceMonitoring:LaunchConfiguration', launchConfiguration_instanceMonitoring - Controls whether instances in this group are launched with detailed (true) or basic (false) monitoring.

For more information, see Configure Monitoring for Auto Scaling Instances in the Amazon EC2 Auto Scaling User Guide.

$sel:metadataOptions:LaunchConfiguration', launchConfiguration_metadataOptions - The metadata options for the instances. For more information, see Configuring the Instance Metadata Options in the Amazon EC2 Auto Scaling User Guide.

$sel:launchConfigurationName:LaunchConfiguration', launchConfiguration_launchConfigurationName - The name of the launch configuration.

$sel:imageId:LaunchConfiguration', launchConfiguration_imageId - The ID of the Amazon Machine Image (AMI) to use to launch your EC2 instances. For more information, see Finding an AMI in the Amazon EC2 User Guide for Linux Instances.

$sel:instanceType:LaunchConfiguration', launchConfiguration_instanceType - The instance type for the instances.

For information about available instance types, see Available Instance Types in the Amazon EC2 User Guide for Linux Instances.

$sel:createdTime:LaunchConfiguration', launchConfiguration_createdTime - The creation date and time for the launch configuration.

launchConfiguration_ebsOptimized :: Lens' LaunchConfiguration (Maybe Bool)

Specifies whether the launch configuration is optimized for EBS I/O (true) or not (false). For more information, see Amazon EBS-Optimized Instances in the Amazon EC2 User Guide for Linux Instances.

launchConfiguration_userData :: Lens' LaunchConfiguration (Maybe Text)

The user data to make available to the launched EC2 instances. For more information, see Instance metadata and user data (Linux) and Instance metadata and user data (Windows). If you are using a command line tool, base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide base64-encoded text. User data is limited to 16 KB.

launchConfiguration_ramdiskId :: Lens' LaunchConfiguration (Maybe Text)

The ID of the RAM disk associated with the AMI.

launchConfiguration_classicLinkVPCSecurityGroups :: Lens' LaunchConfiguration (Maybe [Text])

The IDs of one or more security groups for the VPC specified in ClassicLinkVPCId.

For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

launchConfiguration_spotPrice :: Lens' LaunchConfiguration (Maybe Text)

The maximum hourly price to be paid for any Spot Instance launched to fulfill the request. Spot Instances are launched when the price you specify exceeds the current Spot price. For more information, see Requesting Spot Instances in the Amazon EC2 Auto Scaling User Guide.

launchConfiguration_associatePublicIpAddress :: Lens' LaunchConfiguration (Maybe Bool)

For Auto Scaling groups that are running in a VPC, specifies whether to assign a public IP address to the group's instances. For more information, see Launching Auto Scaling instances in a VPC in the Amazon EC2 Auto Scaling User Guide.

launchConfiguration_securityGroups :: Lens' LaunchConfiguration (Maybe [Text])

A list that contains the security groups to assign to the instances in the Auto Scaling group. For more information, see Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide.

launchConfiguration_iamInstanceProfile :: Lens' LaunchConfiguration (Maybe Text)

The name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role. For more information, see IAM role for applications that run on Amazon EC2 instances in the Amazon EC2 Auto Scaling User Guide.

launchConfiguration_classicLinkVPCId :: Lens' LaunchConfiguration (Maybe Text)

The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. For more information, see ClassicLink in the Amazon EC2 User Guide for Linux Instances and Linking EC2-Classic instances to a VPC in the Amazon EC2 Auto Scaling User Guide.

launchConfiguration_blockDeviceMappings :: Lens' LaunchConfiguration (Maybe [BlockDeviceMapping])

A block device mapping, which specifies the block devices for the instance. For more information, see Block Device Mapping in the Amazon EC2 User Guide for Linux Instances.

launchConfiguration_kernelId :: Lens' LaunchConfiguration (Maybe Text)

The ID of the kernel associated with the AMI.

launchConfiguration_placementTenancy :: Lens' LaunchConfiguration (Maybe Text)

The tenancy of the instance, either default or dedicated. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

For more information, see Configuring instance tenancy with Amazon EC2 Auto Scaling in the Amazon EC2 Auto Scaling User Guide.

launchConfiguration_launchConfigurationARN :: Lens' LaunchConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the launch configuration.

launchConfiguration_keyName :: Lens' LaunchConfiguration (Maybe Text)

The name of the key pair.

For more information, see Amazon EC2 Key Pairs in the Amazon EC2 User Guide for Linux Instances.

launchConfiguration_instanceMonitoring :: Lens' LaunchConfiguration (Maybe InstanceMonitoring)

Controls whether instances in this group are launched with detailed (true) or basic (false) monitoring.

For more information, see Configure Monitoring for Auto Scaling Instances in the Amazon EC2 Auto Scaling User Guide.

launchConfiguration_metadataOptions :: Lens' LaunchConfiguration (Maybe InstanceMetadataOptions)

The metadata options for the instances. For more information, see Configuring the Instance Metadata Options in the Amazon EC2 Auto Scaling User Guide.

launchConfiguration_launchConfigurationName :: Lens' LaunchConfiguration Text

The name of the launch configuration.

launchConfiguration_imageId :: Lens' LaunchConfiguration Text

The ID of the Amazon Machine Image (AMI) to use to launch your EC2 instances. For more information, see Finding an AMI in the Amazon EC2 User Guide for Linux Instances.

launchConfiguration_instanceType :: Lens' LaunchConfiguration Text

The instance type for the instances.

For information about available instance types, see Available Instance Types in the Amazon EC2 User Guide for Linux Instances.

launchConfiguration_createdTime :: Lens' LaunchConfiguration UTCTime

The creation date and time for the launch configuration.