Network.AWS.AutoScaling.Types.LifecycleHook

Description

Documentation

data LifecycleHook

Describes a lifecycle hook, which tells Amazon EC2 Auto Scaling that you want to perform an action whenever it launches instances or terminates instances.

See: newLifecycleHook smart constructor.

Constructors

LifecycleHook'

Fields

defaultResult :: Maybe Text

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The possible values are CONTINUE and ABANDON.

lifecycleHookName :: Maybe Text

The name of the lifecycle hook.

heartbeatTimeout :: Maybe Int

The maximum time, in seconds, that can elapse before the lifecycle hook times out. If the lifecycle hook times out, Amazon EC2 Auto Scaling performs the action that you specified in the DefaultResult parameter.

autoScalingGroupName :: Maybe Text

The name of the Auto Scaling group for the lifecycle hook.

notificationMetadata :: Maybe Text

Additional information that is included any time Amazon EC2 Auto Scaling sends a message to the notification target.

globalTimeout :: Maybe Int

The maximum time, in seconds, that an instance can remain in a Pending:Wait or Terminating:Wait state. The maximum is 172800 seconds (48 hours) or 100 times HeartbeatTimeout, whichever is smaller.

notificationTargetARN :: Maybe Text

The ARN of the target that Amazon EC2 Auto Scaling sends notifications to when an instance is in the transition state for the lifecycle hook. The notification target can be either an SQS queue or an SNS topic.

lifecycleTransition :: Maybe Text

The state of the EC2 instance to which to attach the lifecycle hook. The following are possible values:

autoscaling:EC2_INSTANCE_LAUNCHING

autoscaling:EC2_INSTANCE_TERMINATING

roleARN :: Maybe Text

The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target.

newLifecycleHook :: LifecycleHook

Create a value of LifecycleHook with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultResult:LifecycleHook', lifecycleHook_defaultResult - Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The possible values are CONTINUE and ABANDON.

$sel:lifecycleHookName:LifecycleHook', lifecycleHook_lifecycleHookName - The name of the lifecycle hook.

$sel:heartbeatTimeout:LifecycleHook', lifecycleHook_heartbeatTimeout - The maximum time, in seconds, that can elapse before the lifecycle hook times out. If the lifecycle hook times out, Amazon EC2 Auto Scaling performs the action that you specified in the DefaultResult parameter.

$sel:autoScalingGroupName:LifecycleHook', lifecycleHook_autoScalingGroupName - The name of the Auto Scaling group for the lifecycle hook.

$sel:notificationMetadata:LifecycleHook', lifecycleHook_notificationMetadata - Additional information that is included any time Amazon EC2 Auto Scaling sends a message to the notification target.

$sel:globalTimeout:LifecycleHook', lifecycleHook_globalTimeout - The maximum time, in seconds, that an instance can remain in a Pending:Wait or Terminating:Wait state. The maximum is 172800 seconds (48 hours) or 100 times HeartbeatTimeout, whichever is smaller.

$sel:notificationTargetARN:LifecycleHook', lifecycleHook_notificationTargetARN - The ARN of the target that Amazon EC2 Auto Scaling sends notifications to when an instance is in the transition state for the lifecycle hook. The notification target can be either an SQS queue or an SNS topic.

$sel:lifecycleTransition:LifecycleHook', lifecycleHook_lifecycleTransition - The state of the EC2 instance to which to attach the lifecycle hook. The following are possible values:

autoscaling:EC2_INSTANCE_LAUNCHING

autoscaling:EC2_INSTANCE_TERMINATING

$sel:roleARN:LifecycleHook', lifecycleHook_roleARN - The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target.

lifecycleHook_defaultResult :: Lens' LifecycleHook (Maybe Text)

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The possible values are CONTINUE and ABANDON.

lifecycleHook_lifecycleHookName :: Lens' LifecycleHook (Maybe Text)

The name of the lifecycle hook.

lifecycleHook_heartbeatTimeout :: Lens' LifecycleHook (Maybe Int)

The maximum time, in seconds, that can elapse before the lifecycle hook times out. If the lifecycle hook times out, Amazon EC2 Auto Scaling performs the action that you specified in the DefaultResult parameter.

lifecycleHook_autoScalingGroupName :: Lens' LifecycleHook (Maybe Text)

The name of the Auto Scaling group for the lifecycle hook.

lifecycleHook_notificationMetadata :: Lens' LifecycleHook (Maybe Text)

Additional information that is included any time Amazon EC2 Auto Scaling sends a message to the notification target.

lifecycleHook_globalTimeout :: Lens' LifecycleHook (Maybe Int)

The maximum time, in seconds, that an instance can remain in a Pending:Wait or Terminating:Wait state. The maximum is 172800 seconds (48 hours) or 100 times HeartbeatTimeout, whichever is smaller.

lifecycleHook_notificationTargetARN :: Lens' LifecycleHook (Maybe Text)

The ARN of the target that Amazon EC2 Auto Scaling sends notifications to when an instance is in the transition state for the lifecycle hook. The notification target can be either an SQS queue or an SNS topic.

lifecycleHook_lifecycleTransition :: Lens' LifecycleHook (Maybe Text)

The state of the EC2 instance to which to attach the lifecycle hook. The following are possible values:

autoscaling:EC2_INSTANCE_LAUNCHING

autoscaling:EC2_INSTANCE_TERMINATING

lifecycleHook_roleARN :: Lens' LifecycleHook (Maybe Text)

The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target.