Network.AWS.AutoScaling.Types.LifecycleHookSpecification

Description

Documentation

data LifecycleHookSpecification

Describes information used to specify a lifecycle hook for an Auto Scaling group.

A lifecycle hook tells Amazon EC2 Auto Scaling to perform an action on an instance when the instance launches (before it is put into service) or as the instance terminates (before it is fully terminated).

This step is a part of the procedure for creating a lifecycle hook for an Auto Scaling group:

(Optional) Create a Lambda function and a rule that allows CloudWatch Events to invoke your Lambda function when Amazon EC2 Auto Scaling launches or terminates instances.

(Optional) Create a notification target and an IAM role. The target can be either an Amazon SQS queue or an Amazon SNS topic. The role allows Amazon EC2 Auto Scaling to publish lifecycle notifications to the target.

__Create the lifecycle hook. Specify whether the hook is used when the instances launch or terminate.__

If you need more time, record the lifecycle action heartbeat to keep the instance in a pending state.

If you finish before the timeout period ends, complete the lifecycle action.

For more information, see Amazon EC2 Auto Scaling lifecycle hooks in the Amazon EC2 Auto Scaling User Guide.

See: newLifecycleHookSpecification smart constructor.

Constructors

LifecycleHookSpecification'

Fields

roleARN :: Maybe Text

The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target, for example, an Amazon SNS topic or an Amazon SQS queue.

notificationTargetARN :: Maybe Text

The ARN of the target that Amazon EC2 Auto Scaling sends notifications to when an instance is in the transition state for the lifecycle hook. The notification target can be either an SQS queue or an SNS topic.

heartbeatTimeout :: Maybe Int

The maximum time, in seconds, that can elapse before the lifecycle hook times out.

If the lifecycle hook times out, Amazon EC2 Auto Scaling performs the action that you specified in the DefaultResult parameter. You can prevent the lifecycle hook from timing out by calling RecordLifecycleActionHeartbeat.

notificationMetadata :: Maybe Text

Additional information that you want to include any time Amazon EC2 Auto Scaling sends a message to the notification target.

defaultResult :: Maybe Text

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The valid values are CONTINUE and ABANDON. The default value is ABANDON.

lifecycleHookName :: Text

The name of the lifecycle hook.

lifecycleTransition :: Text

The state of the EC2 instance to which you want to attach the lifecycle hook. The valid values are:

autoscaling:EC2_INSTANCE_LAUNCHING

autoscaling:EC2_INSTANCE_TERMINATING

newLifecycleHookSpecification

Arguments

:: Text
$sel:lifecycleHookName:LifecycleHookSpecification'

-> Text
$sel:lifecycleTransition:LifecycleHookSpecification'

-> LifecycleHookSpecification

Create a value of LifecycleHookSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleARN:LifecycleHookSpecification', lifecycleHookSpecification_roleARN - The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target, for example, an Amazon SNS topic or an Amazon SQS queue.

$sel:notificationTargetARN:LifecycleHookSpecification', lifecycleHookSpecification_notificationTargetARN - The ARN of the target that Amazon EC2 Auto Scaling sends notifications to when an instance is in the transition state for the lifecycle hook. The notification target can be either an SQS queue or an SNS topic.

$sel:heartbeatTimeout:LifecycleHookSpecification', lifecycleHookSpecification_heartbeatTimeout - The maximum time, in seconds, that can elapse before the lifecycle hook times out.

If the lifecycle hook times out, Amazon EC2 Auto Scaling performs the action that you specified in the DefaultResult parameter. You can prevent the lifecycle hook from timing out by calling RecordLifecycleActionHeartbeat.

$sel:notificationMetadata:LifecycleHookSpecification', lifecycleHookSpecification_notificationMetadata - Additional information that you want to include any time Amazon EC2 Auto Scaling sends a message to the notification target.

$sel:defaultResult:LifecycleHookSpecification', lifecycleHookSpecification_defaultResult - Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The valid values are CONTINUE and ABANDON. The default value is ABANDON.

$sel:lifecycleHookName:LifecycleHookSpecification', lifecycleHookSpecification_lifecycleHookName - The name of the lifecycle hook.

$sel:lifecycleTransition:LifecycleHookSpecification', lifecycleHookSpecification_lifecycleTransition - The state of the EC2 instance to which you want to attach the lifecycle hook. The valid values are:

autoscaling:EC2_INSTANCE_LAUNCHING

autoscaling:EC2_INSTANCE_TERMINATING

lifecycleHookSpecification_roleARN :: Lens' LifecycleHookSpecification (Maybe Text)

The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target, for example, an Amazon SNS topic or an Amazon SQS queue.

lifecycleHookSpecification_notificationTargetARN :: Lens' LifecycleHookSpecification (Maybe Text)

The ARN of the target that Amazon EC2 Auto Scaling sends notifications to when an instance is in the transition state for the lifecycle hook. The notification target can be either an SQS queue or an SNS topic.

lifecycleHookSpecification_heartbeatTimeout :: Lens' LifecycleHookSpecification (Maybe Int)

The maximum time, in seconds, that can elapse before the lifecycle hook times out.

If the lifecycle hook times out, Amazon EC2 Auto Scaling performs the action that you specified in the DefaultResult parameter. You can prevent the lifecycle hook from timing out by calling RecordLifecycleActionHeartbeat.

lifecycleHookSpecification_notificationMetadata :: Lens' LifecycleHookSpecification (Maybe Text)

Additional information that you want to include any time Amazon EC2 Auto Scaling sends a message to the notification target.

lifecycleHookSpecification_defaultResult :: Lens' LifecycleHookSpecification (Maybe Text)

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The valid values are CONTINUE and ABANDON. The default value is ABANDON.

lifecycleHookSpecification_lifecycleHookName :: Lens' LifecycleHookSpecification Text

The name of the lifecycle hook.

lifecycleHookSpecification_lifecycleTransition :: Lens' LifecycleHookSpecification Text

The state of the EC2 instance to which you want to attach the lifecycle hook. The valid values are:

autoscaling:EC2_INSTANCE_LAUNCHING

autoscaling:EC2_INSTANCE_TERMINATING