Network.AWS.AutoScaling.Types.MixedInstancesPolicy

Description

Documentation

data MixedInstancesPolicy

Describes a mixed instances policy. A mixed instances policy contains the instance types Amazon EC2 Auto Scaling can launch, and other information Amazon EC2 Auto Scaling can use to launch instances to help you optimize your costs. For more information, see Auto Scaling groups with multiple instance types and purchase options in the Amazon EC2 Auto Scaling User Guide.

See: newMixedInstancesPolicy smart constructor.

Constructors

MixedInstancesPolicy'

Fields

instancesDistribution :: Maybe InstancesDistribution

Specifies the instances distribution. If not provided, the value for each property in InstancesDistribution uses a default value.

launchTemplate :: Maybe LaunchTemplate

Specifies the launch template to use and the instance types (overrides) that are used to provision EC2 instances to fulfill On-Demand and Spot capacities. Required when creating a mixed instances policy.

newMixedInstancesPolicy :: MixedInstancesPolicy

Create a value of MixedInstancesPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instancesDistribution:MixedInstancesPolicy', mixedInstancesPolicy_instancesDistribution - Specifies the instances distribution. If not provided, the value for each property in InstancesDistribution uses a default value.

$sel:launchTemplate:MixedInstancesPolicy', mixedInstancesPolicy_launchTemplate - Specifies the launch template to use and the instance types (overrides) that are used to provision EC2 instances to fulfill On-Demand and Spot capacities. Required when creating a mixed instances policy.

mixedInstancesPolicy_instancesDistribution :: Lens' MixedInstancesPolicy (Maybe InstancesDistribution)

Specifies the instances distribution. If not provided, the value for each property in InstancesDistribution uses a default value.

mixedInstancesPolicy_launchTemplate :: Lens' MixedInstancesPolicy (Maybe LaunchTemplate)

Specifies the launch template to use and the instance types (overrides) that are used to provision EC2 instances to fulfill On-Demand and Spot capacities. Required when creating a mixed instances policy.