Network.AWS.AutoScaling.Types.PredictiveScalingPredefinedScalingMetric

Description

Documentation

data PredictiveScalingPredefinedScalingMetric

Describes a scaling metric for a predictive scaling policy.

When returned in the output of DescribePolicies, it indicates that a predictive scaling policy uses individually specified load and scaling metrics instead of a metric pair.

See: newPredictiveScalingPredefinedScalingMetric smart constructor.

Constructors

PredictiveScalingPredefinedScalingMetric'

Fields

resourceLabel :: Maybe Text

A label that uniquely identifies a specific Application Load Balancer target group from which to determine the average request count served by your Auto Scaling group. You can't specify a resource label unless the target group is attached to the Auto Scaling group.

You create the resource label by appending the final portion of the load balancer ARN and the final portion of the target group ARN into a single value, separated by a forward slash (/). The format of the resource label is:

app/my-alb/778d41231b141a0f/targetgroup/my-alb-target-group/943f017f100becff.

Where:

app/<load-balancer-name>/<load-balancer-id> is the final portion of the load balancer ARN

targetgroup/<target-group-name>/<target-group-id> is the final portion of the target group ARN.

To find the ARN for an Application Load Balancer, use the DescribeLoadBalancers API operation. To find the ARN for the target group, use the DescribeTargetGroups API operation.

predefinedMetricType :: PredefinedScalingMetricType

The metric type.

newPredictiveScalingPredefinedScalingMetric

Arguments

:: PredefinedScalingMetricType
$sel:predefinedMetricType:PredictiveScalingPredefinedScalingMetric'

-> PredictiveScalingPredefinedScalingMetric

Create a value of PredictiveScalingPredefinedScalingMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceLabel:PredictiveScalingPredefinedScalingMetric', predictiveScalingPredefinedScalingMetric_resourceLabel - A label that uniquely identifies a specific Application Load Balancer target group from which to determine the average request count served by your Auto Scaling group. You can't specify a resource label unless the target group is attached to the Auto Scaling group.

You create the resource label by appending the final portion of the load balancer ARN and the final portion of the target group ARN into a single value, separated by a forward slash (/). The format of the resource label is:

app/my-alb/778d41231b141a0f/targetgroup/my-alb-target-group/943f017f100becff.

Where:

app/<load-balancer-name>/<load-balancer-id> is the final portion of the load balancer ARN

targetgroup/<target-group-name>/<target-group-id> is the final portion of the target group ARN.

To find the ARN for an Application Load Balancer, use the DescribeLoadBalancers API operation. To find the ARN for the target group, use the DescribeTargetGroups API operation.

$sel:predefinedMetricType:PredictiveScalingPredefinedScalingMetric', predictiveScalingPredefinedScalingMetric_predefinedMetricType - The metric type.

predictiveScalingPredefinedScalingMetric_resourceLabel :: Lens' PredictiveScalingPredefinedScalingMetric (Maybe Text)

A label that uniquely identifies a specific Application Load Balancer target group from which to determine the average request count served by your Auto Scaling group. You can't specify a resource label unless the target group is attached to the Auto Scaling group.

You create the resource label by appending the final portion of the load balancer ARN and the final portion of the target group ARN into a single value, separated by a forward slash (/). The format of the resource label is:

app/my-alb/778d41231b141a0f/targetgroup/my-alb-target-group/943f017f100becff.

Where:

app/<load-balancer-name>/<load-balancer-id> is the final portion of the load balancer ARN

targetgroup/<target-group-name>/<target-group-id> is the final portion of the target group ARN.

To find the ARN for an Application Load Balancer, use the DescribeLoadBalancers API operation. To find the ARN for the target group, use the DescribeTargetGroups API operation.

predictiveScalingPredefinedScalingMetric_predefinedMetricType :: Lens' PredictiveScalingPredefinedScalingMetric PredefinedScalingMetricType

The metric type.