Network.AWS.AutoScaling.Types.ScalingPolicy

Description

Documentation

data ScalingPolicy

Describes a scaling policy.

See: newScalingPolicy smart constructor.

Constructors

ScalingPolicy'

Fields

minAdjustmentStep :: Maybe Int

Available for backward compatibility. Use MinAdjustmentMagnitude instead.

estimatedInstanceWarmup :: Maybe Int

The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics.

policyName :: Maybe Text

The name of the scaling policy.

enabled :: Maybe Bool

Indicates whether the policy is enabled (true) or disabled (false).

policyType :: Maybe Text

One of the following policy types:

TargetTrackingScaling

StepScaling

SimpleScaling (default)

PredictiveScaling

For more information, see Target tracking scaling policies and Step and simple scaling policies in the Amazon EC2 Auto Scaling User Guide.

stepAdjustments :: Maybe [StepAdjustment]

A set of adjustments that enable you to scale based on the size of the alarm breach.

targetTrackingConfiguration :: Maybe TargetTrackingConfiguration

A target tracking scaling policy.

adjustmentType :: Maybe Text

Specifies how the scaling adjustment is interpreted (for example, an absolute number or a percentage). The valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

autoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

predictiveScalingConfiguration :: Maybe PredictiveScalingConfiguration

A predictive scaling policy.

scalingAdjustment :: Maybe Int

The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity.

cooldown :: Maybe Int

The duration of the policy's cooldown period, in seconds.

policyARN :: Maybe Text

The Amazon Resource Name (ARN) of the policy.

alarms :: Maybe [Alarm]

The CloudWatch alarms related to the policy.

metricAggregationType :: Maybe Text

The aggregation type for the CloudWatch metrics. The valid values are Minimum, Maximum, and Average.

minAdjustmentMagnitude :: Maybe Int

The minimum value to scale by when the adjustment type is PercentChangeInCapacity.

newScalingPolicy :: ScalingPolicy

Create a value of ScalingPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minAdjustmentStep:ScalingPolicy', scalingPolicy_minAdjustmentStep - Available for backward compatibility. Use MinAdjustmentMagnitude instead.

$sel:estimatedInstanceWarmup:ScalingPolicy', scalingPolicy_estimatedInstanceWarmup - The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics.

$sel:policyName:ScalingPolicy', scalingPolicy_policyName - The name of the scaling policy.

$sel:enabled:ScalingPolicy', scalingPolicy_enabled - Indicates whether the policy is enabled (true) or disabled (false).

$sel:policyType:ScalingPolicy', scalingPolicy_policyType - One of the following policy types:

TargetTrackingScaling

StepScaling

SimpleScaling (default)

PredictiveScaling

For more information, see Target tracking scaling policies and Step and simple scaling policies in the Amazon EC2 Auto Scaling User Guide.

$sel:stepAdjustments:ScalingPolicy', scalingPolicy_stepAdjustments - A set of adjustments that enable you to scale based on the size of the alarm breach.

$sel:targetTrackingConfiguration:ScalingPolicy', scalingPolicy_targetTrackingConfiguration - A target tracking scaling policy.

$sel:adjustmentType:ScalingPolicy', scalingPolicy_adjustmentType - Specifies how the scaling adjustment is interpreted (for example, an absolute number or a percentage). The valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

$sel:autoScalingGroupName:ScalingPolicy', scalingPolicy_autoScalingGroupName - The name of the Auto Scaling group.

$sel:predictiveScalingConfiguration:ScalingPolicy', scalingPolicy_predictiveScalingConfiguration - A predictive scaling policy.

$sel:scalingAdjustment:ScalingPolicy', scalingPolicy_scalingAdjustment - The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity.

$sel:cooldown:ScalingPolicy', scalingPolicy_cooldown - The duration of the policy's cooldown period, in seconds.

$sel:policyARN:ScalingPolicy', scalingPolicy_policyARN - The Amazon Resource Name (ARN) of the policy.

$sel:alarms:ScalingPolicy', scalingPolicy_alarms - The CloudWatch alarms related to the policy.

$sel:metricAggregationType:ScalingPolicy', scalingPolicy_metricAggregationType - The aggregation type for the CloudWatch metrics. The valid values are Minimum, Maximum, and Average.

$sel:minAdjustmentMagnitude:ScalingPolicy', scalingPolicy_minAdjustmentMagnitude - The minimum value to scale by when the adjustment type is PercentChangeInCapacity.

scalingPolicy_minAdjustmentStep :: Lens' ScalingPolicy (Maybe Int)

Available for backward compatibility. Use MinAdjustmentMagnitude instead.

scalingPolicy_estimatedInstanceWarmup :: Lens' ScalingPolicy (Maybe Int)

The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics.

scalingPolicy_policyName :: Lens' ScalingPolicy (Maybe Text)

The name of the scaling policy.

scalingPolicy_enabled :: Lens' ScalingPolicy (Maybe Bool)

Indicates whether the policy is enabled (true) or disabled (false).

scalingPolicy_policyType :: Lens' ScalingPolicy (Maybe Text)

One of the following policy types:

TargetTrackingScaling

StepScaling

SimpleScaling (default)

PredictiveScaling

For more information, see Target tracking scaling policies and Step and simple scaling policies in the Amazon EC2 Auto Scaling User Guide.

scalingPolicy_stepAdjustments :: Lens' ScalingPolicy (Maybe [StepAdjustment])

A set of adjustments that enable you to scale based on the size of the alarm breach.

scalingPolicy_targetTrackingConfiguration :: Lens' ScalingPolicy (Maybe TargetTrackingConfiguration)

A target tracking scaling policy.

scalingPolicy_adjustmentType :: Lens' ScalingPolicy (Maybe Text)

Specifies how the scaling adjustment is interpreted (for example, an absolute number or a percentage). The valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

scalingPolicy_autoScalingGroupName :: Lens' ScalingPolicy (Maybe Text)

The name of the Auto Scaling group.

scalingPolicy_predictiveScalingConfiguration :: Lens' ScalingPolicy (Maybe PredictiveScalingConfiguration)

A predictive scaling policy.

scalingPolicy_scalingAdjustment :: Lens' ScalingPolicy (Maybe Int)

The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity.

scalingPolicy_cooldown :: Lens' ScalingPolicy (Maybe Int)

The duration of the policy's cooldown period, in seconds.

scalingPolicy_policyARN :: Lens' ScalingPolicy (Maybe Text)

The Amazon Resource Name (ARN) of the policy.

scalingPolicy_alarms :: Lens' ScalingPolicy (Maybe [Alarm])

The CloudWatch alarms related to the policy.

scalingPolicy_metricAggregationType :: Lens' ScalingPolicy (Maybe Text)

The aggregation type for the CloudWatch metrics. The valid values are Minimum, Maximum, and Average.

scalingPolicy_minAdjustmentMagnitude :: Lens' ScalingPolicy (Maybe Int)

The minimum value to scale by when the adjustment type is PercentChangeInCapacity.