Network.AWS.AutoScaling.Types.ScheduledUpdateGroupAction

Description

Documentation

data ScheduledUpdateGroupAction

Describes a scheduled scaling action.

See: newScheduledUpdateGroupAction smart constructor.

Constructors

ScheduledUpdateGroupAction'

Fields

minSize :: Maybe Int

The minimum size of the Auto Scaling group.

desiredCapacity :: Maybe Int

The desired capacity is the initial capacity of the Auto Scaling group after the scheduled action runs and the capacity it attempts to maintain.

startTime :: Maybe ISO8601

The date and time in UTC for this action to start. For example, "2019-06-01T00:00:00Z".

endTime :: Maybe ISO8601

The date and time in UTC for the recurring schedule to end. For example, "2019-06-01T00:00:00Z".

scheduledActionARN :: Maybe Text

The Amazon Resource Name (ARN) of the scheduled action.

recurrence :: Maybe Text

The recurring schedule for the action, in Unix cron syntax format.

When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action starts and stops.

maxSize :: Maybe Int

The maximum size of the Auto Scaling group.

timeZone :: Maybe Text

The time zone for the cron expression.

scheduledActionName :: Maybe Text

The name of the scheduled action.

time :: Maybe ISO8601

This parameter is no longer used.

autoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

newScheduledUpdateGroupAction :: ScheduledUpdateGroupAction

Create a value of ScheduledUpdateGroupAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minSize:ScheduledUpdateGroupAction', scheduledUpdateGroupAction_minSize - The minimum size of the Auto Scaling group.

$sel:desiredCapacity:ScheduledUpdateGroupAction', scheduledUpdateGroupAction_desiredCapacity - The desired capacity is the initial capacity of the Auto Scaling group after the scheduled action runs and the capacity it attempts to maintain.

$sel:startTime:ScheduledUpdateGroupAction', scheduledUpdateGroupAction_startTime - The date and time in UTC for this action to start. For example, "2019-06-01T00:00:00Z".

$sel:endTime:ScheduledUpdateGroupAction', scheduledUpdateGroupAction_endTime - The date and time in UTC for the recurring schedule to end. For example, "2019-06-01T00:00:00Z".

$sel:scheduledActionARN:ScheduledUpdateGroupAction', scheduledUpdateGroupAction_scheduledActionARN - The Amazon Resource Name (ARN) of the scheduled action.

$sel:recurrence:ScheduledUpdateGroupAction', scheduledUpdateGroupAction_recurrence - The recurring schedule for the action, in Unix cron syntax format.

When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action starts and stops.

$sel:maxSize:ScheduledUpdateGroupAction', scheduledUpdateGroupAction_maxSize - The maximum size of the Auto Scaling group.

$sel:timeZone:ScheduledUpdateGroupAction', scheduledUpdateGroupAction_timeZone - The time zone for the cron expression.

$sel:scheduledActionName:ScheduledUpdateGroupAction', scheduledUpdateGroupAction_scheduledActionName - The name of the scheduled action.

$sel:time:ScheduledUpdateGroupAction', scheduledUpdateGroupAction_time - This parameter is no longer used.

$sel:autoScalingGroupName:ScheduledUpdateGroupAction', scheduledUpdateGroupAction_autoScalingGroupName - The name of the Auto Scaling group.

scheduledUpdateGroupAction_minSize :: Lens' ScheduledUpdateGroupAction (Maybe Int)

The minimum size of the Auto Scaling group.

scheduledUpdateGroupAction_desiredCapacity :: Lens' ScheduledUpdateGroupAction (Maybe Int)

The desired capacity is the initial capacity of the Auto Scaling group after the scheduled action runs and the capacity it attempts to maintain.

scheduledUpdateGroupAction_startTime :: Lens' ScheduledUpdateGroupAction (Maybe UTCTime)

The date and time in UTC for this action to start. For example, "2019-06-01T00:00:00Z".

scheduledUpdateGroupAction_endTime :: Lens' ScheduledUpdateGroupAction (Maybe UTCTime)

The date and time in UTC for the recurring schedule to end. For example, "2019-06-01T00:00:00Z".

scheduledUpdateGroupAction_scheduledActionARN :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The Amazon Resource Name (ARN) of the scheduled action.

scheduledUpdateGroupAction_recurrence :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The recurring schedule for the action, in Unix cron syntax format.

When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action starts and stops.

scheduledUpdateGroupAction_maxSize :: Lens' ScheduledUpdateGroupAction (Maybe Int)

The maximum size of the Auto Scaling group.

scheduledUpdateGroupAction_timeZone :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The time zone for the cron expression.

scheduledUpdateGroupAction_scheduledActionName :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The name of the scheduled action.

scheduledUpdateGroupAction_time :: Lens' ScheduledUpdateGroupAction (Maybe UTCTime)

This parameter is no longer used.

scheduledUpdateGroupAction_autoScalingGroupName :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The name of the Auto Scaling group.