Network.AWS.AutoScaling.Types.ScheduledUpdateGroupActionRequest

Description

Documentation

data ScheduledUpdateGroupActionRequest

Describes information used for one or more scheduled scaling action updates in a BatchPutScheduledUpdateGroupAction operation.

See: newScheduledUpdateGroupActionRequest smart constructor.

Constructors

ScheduledUpdateGroupActionRequest'

Fields

startTime :: Maybe ISO8601

The date and time for the action to start, in YYYY-MM-DDThh:mm:ssZ format in UTC/GMT only and in quotes (for example, "2019-06-01T00:00:00Z").

If you specify Recurrence and StartTime, Amazon EC2 Auto Scaling performs the action at this time, and then performs the action based on the specified recurrence.

If you try to schedule the action in the past, Amazon EC2 Auto Scaling returns an error message.

maxSize :: Maybe Int

The maximum size of the Auto Scaling group.

recurrence :: Maybe Text

The recurring schedule for the action, in Unix cron syntax format. This format consists of five fields separated by white spaces: [Minute] [Hour] [Day_of_Month] [Month_of_Year] [Day_of_Week]. The value must be in quotes (for example, "30 0 1 1,6,12 *"). For more information about this format, see Crontab.

When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action starts and stops.

Cron expressions use Universal Coordinated Time (UTC) by default.

desiredCapacity :: Maybe Int

The desired capacity is the initial capacity of the Auto Scaling group after the scheduled action runs and the capacity it attempts to maintain.

minSize :: Maybe Int

The minimum size of the Auto Scaling group.

endTime :: Maybe ISO8601

The date and time for the recurring schedule to end, in UTC.

timeZone :: Maybe Text

Specifies the time zone for a cron expression. If a time zone is not provided, UTC is used by default.

Valid values are the canonical names of the IANA time zones, derived from the IANA Time Zone Database (such as Etc/GMT+9 or Pacific/Tahiti). For more information, see https://en.wikipedia.org/wiki/List_of_tz_database_time_zones.

scheduledActionName :: Text

The name of the scaling action.

newScheduledUpdateGroupActionRequest

Arguments

:: Text
$sel:scheduledActionName:ScheduledUpdateGroupActionRequest'

-> ScheduledUpdateGroupActionRequest

Create a value of ScheduledUpdateGroupActionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:ScheduledUpdateGroupActionRequest', scheduledUpdateGroupActionRequest_startTime - The date and time for the action to start, in YYYY-MM-DDThh:mm:ssZ format in UTC/GMT only and in quotes (for example, "2019-06-01T00:00:00Z").

If you specify Recurrence and StartTime, Amazon EC2 Auto Scaling performs the action at this time, and then performs the action based on the specified recurrence.

If you try to schedule the action in the past, Amazon EC2 Auto Scaling returns an error message.

$sel:maxSize:ScheduledUpdateGroupActionRequest', scheduledUpdateGroupActionRequest_maxSize - The maximum size of the Auto Scaling group.

$sel:recurrence:ScheduledUpdateGroupActionRequest', scheduledUpdateGroupActionRequest_recurrence - The recurring schedule for the action, in Unix cron syntax format. This format consists of five fields separated by white spaces: [Minute] [Hour] [Day_of_Month] [Month_of_Year] [Day_of_Week]. The value must be in quotes (for example, "30 0 1 1,6,12 *"). For more information about this format, see Crontab.

When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action starts and stops.

Cron expressions use Universal Coordinated Time (UTC) by default.

$sel:desiredCapacity:ScheduledUpdateGroupActionRequest', scheduledUpdateGroupActionRequest_desiredCapacity - The desired capacity is the initial capacity of the Auto Scaling group after the scheduled action runs and the capacity it attempts to maintain.

$sel:minSize:ScheduledUpdateGroupActionRequest', scheduledUpdateGroupActionRequest_minSize - The minimum size of the Auto Scaling group.

$sel:endTime:ScheduledUpdateGroupActionRequest', scheduledUpdateGroupActionRequest_endTime - The date and time for the recurring schedule to end, in UTC.

$sel:timeZone:ScheduledUpdateGroupActionRequest', scheduledUpdateGroupActionRequest_timeZone - Specifies the time zone for a cron expression. If a time zone is not provided, UTC is used by default.

Valid values are the canonical names of the IANA time zones, derived from the IANA Time Zone Database (such as Etc/GMT+9 or Pacific/Tahiti). For more information, see https://en.wikipedia.org/wiki/List_of_tz_database_time_zones.

$sel:scheduledActionName:ScheduledUpdateGroupActionRequest', scheduledUpdateGroupActionRequest_scheduledActionName - The name of the scaling action.

scheduledUpdateGroupActionRequest_startTime :: Lens' ScheduledUpdateGroupActionRequest (Maybe UTCTime)

The date and time for the action to start, in YYYY-MM-DDThh:mm:ssZ format in UTC/GMT only and in quotes (for example, "2019-06-01T00:00:00Z").

If you specify Recurrence and StartTime, Amazon EC2 Auto Scaling performs the action at this time, and then performs the action based on the specified recurrence.

If you try to schedule the action in the past, Amazon EC2 Auto Scaling returns an error message.

scheduledUpdateGroupActionRequest_maxSize :: Lens' ScheduledUpdateGroupActionRequest (Maybe Int)

The maximum size of the Auto Scaling group.

scheduledUpdateGroupActionRequest_recurrence :: Lens' ScheduledUpdateGroupActionRequest (Maybe Text)

The recurring schedule for the action, in Unix cron syntax format. This format consists of five fields separated by white spaces: [Minute] [Hour] [Day_of_Month] [Month_of_Year] [Day_of_Week]. The value must be in quotes (for example, "30 0 1 1,6,12 *"). For more information about this format, see Crontab.

When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action starts and stops.

Cron expressions use Universal Coordinated Time (UTC) by default.

scheduledUpdateGroupActionRequest_desiredCapacity :: Lens' ScheduledUpdateGroupActionRequest (Maybe Int)

The desired capacity is the initial capacity of the Auto Scaling group after the scheduled action runs and the capacity it attempts to maintain.

scheduledUpdateGroupActionRequest_minSize :: Lens' ScheduledUpdateGroupActionRequest (Maybe Int)

The minimum size of the Auto Scaling group.

scheduledUpdateGroupActionRequest_endTime :: Lens' ScheduledUpdateGroupActionRequest (Maybe UTCTime)

The date and time for the recurring schedule to end, in UTC.

scheduledUpdateGroupActionRequest_timeZone :: Lens' ScheduledUpdateGroupActionRequest (Maybe Text)

Specifies the time zone for a cron expression. If a time zone is not provided, UTC is used by default.

Valid values are the canonical names of the IANA time zones, derived from the IANA Time Zone Database (such as Etc/GMT+9 or Pacific/Tahiti). For more information, see https://en.wikipedia.org/wiki/List_of_tz_database_time_zones.

scheduledUpdateGroupActionRequest_scheduledActionName :: Lens' ScheduledUpdateGroupActionRequest Text

The name of the scaling action.