Network.AWS.AutoScaling.Types.WarmPoolConfiguration

Description

Documentation

data WarmPoolConfiguration

Describes a warm pool configuration.

See: newWarmPoolConfiguration smart constructor.

Constructors

WarmPoolConfiguration'

Fields

minSize :: Maybe Natural

The minimum number of instances to maintain in the warm pool.

status :: Maybe WarmPoolStatus

The status of a warm pool that is marked for deletion.

poolState :: Maybe WarmPoolState

The instance state to transition to after the lifecycle actions are complete.

maxGroupPreparedCapacity :: Maybe Int

The maximum number of instances that are allowed to be in the warm pool or in any state except Terminated for the Auto Scaling group.

newWarmPoolConfiguration :: WarmPoolConfiguration

Create a value of WarmPoolConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minSize:WarmPoolConfiguration', warmPoolConfiguration_minSize - The minimum number of instances to maintain in the warm pool.

$sel:status:WarmPoolConfiguration', warmPoolConfiguration_status - The status of a warm pool that is marked for deletion.

$sel:poolState:WarmPoolConfiguration', warmPoolConfiguration_poolState - The instance state to transition to after the lifecycle actions are complete.

$sel:maxGroupPreparedCapacity:WarmPoolConfiguration', warmPoolConfiguration_maxGroupPreparedCapacity - The maximum number of instances that are allowed to be in the warm pool or in any state except Terminated for the Auto Scaling group.

warmPoolConfiguration_minSize :: Lens' WarmPoolConfiguration (Maybe Natural)

The minimum number of instances to maintain in the warm pool.

warmPoolConfiguration_status :: Lens' WarmPoolConfiguration (Maybe WarmPoolStatus)

The status of a warm pool that is marked for deletion.

warmPoolConfiguration_poolState :: Lens' WarmPoolConfiguration (Maybe WarmPoolState)

The instance state to transition to after the lifecycle actions are complete.

warmPoolConfiguration_maxGroupPreparedCapacity :: Lens' WarmPoolConfiguration (Maybe Int)

The maximum number of instances that are allowed to be in the warm pool or in any state except Terminated for the Auto Scaling group.