Network.AWS.AutoScaling.V2011_01_01.AttachInstances

Description

Attaches one or more Amazon EC2 instances to an existing Auto Scaling group. After the instance(s) is attached, it becomes a part of the Auto Scaling group. For more information, see Attach Amazon EC2 Instance(s) to Your Existing Auto Scaling Group in the Auto Scaling Developer Guide.

Documentation

attachInstances

Arguments

:: Text
_aiqAutoScalingGroupName

-> AttachInstances

Minimum specification for a AttachInstances request.

data AttachInstances

Constructors

AttachInstances

Fields

_aiqAutoScalingGroupName :: Text

The name of the Auto Scaling group to which to attach the specified instance(s).

_aiqInstanceIds :: [Text]

One or more IDs of the Amazon EC2 instances to attach to the specified Auto Scaling group. You must specify at least one instance ID.

aiqInstanceIds :: Lens' AttachInstances [Text]

aiqAutoScalingGroupName :: Lens' AttachInstances Text

data AttachInstancesResponse

Constructors

AttachInstancesResponse