Network.AWS.AutoScaling.V2011_01_01.DeleteScheduledAction

Description

Deletes a scheduled action previously created using the PutScheduledUpdateGroupAction.

Documentation

deleteScheduledAction

Arguments

:: Text
_dsatScheduledActionName

-> DeleteScheduledAction

Minimum specification for a DeleteScheduledAction request.

data DeleteScheduledAction

Constructors

DeleteScheduledAction

Fields

_dsatScheduledActionName :: Text

The name of the action you want to delete.

_dsatAutoScalingGroupName :: Maybe Text

The name of the Auto Scaling group.

dsatScheduledActionName :: Lens' DeleteScheduledAction Text

dsatAutoScalingGroupName :: Lens' DeleteScheduledAction (Maybe Text)

data DeleteScheduledActionResponse

Constructors

DeleteScheduledActionResponse