Network.AWS.AutoScaling.V2011_01_01.DescribeAutoScalingNotificationTypes

Description

Returns a list of all notification types that are supported by Auto Scaling.

Documentation

data DescribeAutoScalingNotificationTypes

Constructors

DescribeAutoScalingNotificationTypes

data DescribeAutoScalingNotificationTypesResponse

Constructors

DescribeAutoScalingNotificationTypesResponse

Fields

_dasntaAutoScalingNotificationTypes :: [Text]

Returns a list of all notification types supported by Auto Scaling. They are: autoscaling:EC2_INSTANCE_LAUNCH autoscaling:EC2_INSTANCE_LAUNCH_ERROR autoscaling:EC2_INSTANCE_TERMINATE autoscaling:EC2_INSTANCE_TERMINATE_ERROR autoscaling:TEST_NOTIFICATION https:/autoscaling.amazonaws.com?Version=2011-01-01&Action=DescribeAutoScalingNotificationTypes &AUTHPARAMS autoscaling:EC2_INSTANCE_LAUNCH autoscaling:EC2_INSTANCE_LAUNCH_ERROR autoscaling:EC2_INSTANCE_TERMINATE autoscaling:EC2_INSTANCE_TERMINATE_ERROR autoscaling:TEST_NOTIFICATION 42fc6794-bf21-11e2-a1cf-ff3dEXAMPLE.

dasntaAutoScalingNotificationTypes :: Iso' DescribeAutoScalingNotificationTypesResponse [Text]