Network.AWS.AutoScaling.V2011_01_01.DescribeLaunchConfigurations

Description

Returns a full description of the launch configurations, or the specified launch configurations, if they exist. If no name is specified, then the full details of all launch configurations are returned. https:/autoscaling.amazonaws.com?LaunchConfigurationNames.member.1=my-test-lc &MaxRecords=20 &Version=2011-01-01 &Action=DescribeLaunchConfigurations &AUTHPARAMS true dedicated 2013-01-21T23:04:42.200Z my-test-lc m1.small arn:aws:autoscaling:us-east-1:803981987763:launchConfiguration: 9dbbbf87-6141-428a-a409-0752edbe6cad:launchConfigurationName/my-test-lc ami-514ac838 true false d05a22f8-b690-11e2-bf8e-2113fEXAMPLE.

Documentation

describeLaunchConfigurations :: DescribeLaunchConfigurations

Minimum specification for a DescribeLaunchConfigurations request.

data DescribeLaunchConfigurations

Constructors

DescribeLaunchConfigurations

Fields

_lcnuLaunchConfigurationNames :: [Text]

A list of launch configuration names.

_lcnuMaxRecords :: Maybe Integer

The maximum number of launch configurations. The default is 100.

_lcnuNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

lcnuNextToken :: Lens' DescribeLaunchConfigurations (Maybe Text)

lcnuMaxRecords :: Lens' DescribeLaunchConfigurations (Maybe Integer)

lcnuLaunchConfigurationNames :: Lens' DescribeLaunchConfigurations [Text]

data DescribeLaunchConfigurationsResponse

Constructors

DescribeLaunchConfigurationsResponse

Fields

_lctLaunchConfigurations :: [LaunchConfiguration]

A list of launch configurations.

_lctNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

lctNextToken :: Lens' DescribeLaunchConfigurationsResponse (Maybe Text)

lctLaunchConfigurations :: Lens' DescribeLaunchConfigurationsResponse [LaunchConfiguration]