Network.AWS.AutoScaling.V2011_01_01.DescribeNotificationConfigurations

Description

Returns a list of notification actions associated with Auto Scaling groups for specified events.

Documentation

describeNotificationConfigurations :: DescribeNotificationConfigurations

Minimum specification for a DescribeNotificationConfigurations request.

data DescribeNotificationConfigurations

Constructors

DescribeNotificationConfigurations

Fields

_dncuAutoScalingGroupNames :: [Text]

The name of the Auto Scaling group.

_dncuMaxRecords :: Maybe Integer

Maximum number of records to be returned.

_dncuNextToken :: Maybe Text

A string that is used to mark the start of the next batch of returned results for pagination.

dncuNextToken :: Lens' DescribeNotificationConfigurations (Maybe Text)

dncuMaxRecords :: Lens' DescribeNotificationConfigurations (Maybe Integer)

dncuAutoScalingGroupNames :: Lens' DescribeNotificationConfigurations [Text]

data DescribeNotificationConfigurationsResponse

Constructors

DescribeNotificationConfigurationsResponse

Fields

_dncaNotificationConfigurations :: [NotificationConfiguration]

The list of notification configurations.

_dncaNextToken :: Maybe Text

A string that is used to mark the start of the next batch of returned results for pagination.

dncaNotificationConfigurations :: Lens' DescribeNotificationConfigurationsResponse [NotificationConfiguration]

dncaNextToken :: Lens' DescribeNotificationConfigurationsResponse (Maybe Text)