Network.AWS.AutoScaling.V2011_01_01.DescribeTags

Description

Lists the Auto Scaling group tags. You can use filters to limit results when describing tags. For example, you can query for tags of a particular Auto Scaling group. You can specify multiple values for a filter. A tag must match at least one of the specified values for it to be included in the results. You can also specify multiple filters. The result includes information for a particular tag only if it matches all your filters. If there's no match, no special message is returned. https:/autoscaling.amazonaws.com?Version=2011-01-01&Action=DescribeTags &AUTHPARAMS my-test-asg true 1.0 version auto-scaling-group 086265fd-bf3e-11e2-85fc-fbb1EXAMPLE.

Documentation

describeTags :: DescribeTags

Minimum specification for a DescribeTags request.

data DescribeTags

Constructors

DescribeTags

Fields

_dttFilters :: [Filter]

The value of the filter type used to identify the tags to be returned. For example, you can filter so that tags are returned according to Auto Scaling group, the key and value, or whether the new tag will be applied to instances launched after the tag is created (PropagateAtLaunch).

_dttMaxRecords :: Maybe Integer

The maximum number of records to return.

_dttNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

dttNextToken :: Lens' DescribeTags (Maybe Text)

dttMaxRecords :: Lens' DescribeTags (Maybe Integer)

dttFilters :: Lens' DescribeTags [Filter]

data DescribeTagsResponse

Constructors

DescribeTagsResponse

Fields

_ttTags :: [TagDescription]

The list of tags.

_ttNextToken :: Maybe Text

A string used to mark the start of the next batch of returned results.

ttTags :: Lens' DescribeTagsResponse [TagDescription]

ttNextToken :: Lens' DescribeTagsResponse (Maybe Text)