Network.AWS.AutoScaling.V2011_01_01.PutScheduledUpdateGroupAction

Description

Creates or updates a scheduled scaling action for an Auto Scaling group. When updating a scheduled scaling action, if you leave a parameter unspecified, the corresponding value remains unchanged in the affected Auto Scaling group. For information on creating or updating a scheduled action for your Auto Scaling group, see Scale Based on a Schedule. Auto Scaling supports the date and time expressed in YYYY-MM-DDThh:mm:ssZ format in UTC/GMT only. Schedule based on a specific date and time https:/autoscaling.amazonaws.com?AutoScalingGroupName=my-test-asg &ScheduledActionName=ScaleUp &StartTime=2013-05-25T08:00:00Z &DesiredCapacity=3 &Version=2011-01-01 &Action=PutScheduledUpdateGroupAction &AUTHPARAMS 3bc8c9bc-6a62-11e2-8a51-4b8a1EXAMPLE Recurring Schedule https:/autoscaling.amazonaws.com?AutoScalingGroupName=my-test-asg &ScheduledActionName=scaleup-schedule-year &Recurrence="30 0 1 1,6,12 *" &DesiredCapacity=3 &Version=2011-01-01 &Action=PutScheduledUpdateGroupAction &AUTHPARAMS 3bc8c9bc-6a62-11e2-8a51-4b8a1EXAMPLE.

Documentation

putScheduledUpdateGroupAction

Arguments

:: Text
_psugatAutoScalingGroupName

-> Text
_psugatScheduledActionName

-> PutScheduledUpdateGroupAction

Minimum specification for a PutScheduledUpdateGroupAction request.

data PutScheduledUpdateGroupAction

Constructors

PutScheduledUpdateGroupAction

Fields

_psugatAutoScalingGroupName :: Text

The name or ARN of the Auto Scaling group.

_psugatScheduledActionName :: Text

The name of this scaling action.

_psugatDesiredCapacity :: Maybe Integer

The number of Amazon EC2 instances that should be running in the group.

_psugatMaxSize :: Maybe Integer

The maximum size for the Auto Scaling group.

_psugatMinSize :: Maybe Integer

The minimum size for the new Auto Scaling group.

_psugatStartTime :: Maybe ISO8601

The time for this action to start, as in --start-time 2010-06-01T00:00:00Z. If you try to schedule your action in the past, Auto Scaling returns an error message. When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action will start and stop.

_psugatTime :: Maybe ISO8601

Time is deprecated. The time for this action to start. Time is an alias for StartTime and can be specified instead of StartTime, or vice versa. If both Time and StartTime are specified, their values should be identical. Otherwise, PutScheduledUpdateGroupAction will return an error.

_psugatEndTime :: Maybe ISO8601

The time for this action to end.

_psugatRecurrence :: Maybe Text

The time when recurring future actions will start. Start time is specified by the user following the Unix cron syntax format. For information about cron syntax, go to Wikipedia, The Free Encyclopedia. When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action will start and stop.

psugatTime :: Lens' PutScheduledUpdateGroupAction (Maybe ISO8601)

psugatStartTime :: Lens' PutScheduledUpdateGroupAction (Maybe ISO8601)

psugatScheduledActionName :: Lens' PutScheduledUpdateGroupAction Text

psugatRecurrence :: Lens' PutScheduledUpdateGroupAction (Maybe Text)

psugatMinSize :: Lens' PutScheduledUpdateGroupAction (Maybe Integer)

psugatMaxSize :: Lens' PutScheduledUpdateGroupAction (Maybe Integer)

psugatEndTime :: Lens' PutScheduledUpdateGroupAction (Maybe ISO8601)

psugatDesiredCapacity :: Lens' PutScheduledUpdateGroupAction (Maybe Integer)

psugatAutoScalingGroupName :: Lens' PutScheduledUpdateGroupAction Text

data PutScheduledUpdateGroupActionResponse

Constructors

PutScheduledUpdateGroupActionResponse