Network.AWS.AutoScaling.V2011_01_01.ResumeProcesses

Description

Resumes all suspended Auto Scaling processes for an Auto Scaling group. For information on suspending and resuming Auto Scaling process, see Suspend and Resume Auto Scaling Process.

Documentation

resumeProcesses

Arguments

:: Text
_sprAutoScalingGroupName

-> ResumeProcesses

Minimum specification for a ResumeProcesses request.

data ResumeProcesses

Constructors

ResumeProcesses

Fields

_sprAutoScalingGroupName :: Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

_sprScalingProcesses :: [Text]

The processes that you want to suspend or resume, which can include one or more of the following: Launch Terminate HealthCheck ReplaceUnhealthy AZRebalance AlarmNotification ScheduledActions AddToLoadBalancer To suspend all process types, omit this parameter.

sprScalingProcesses :: Lens' ResumeProcesses [Text]

sprAutoScalingGroupName :: Lens' ResumeProcesses Text

data ResumeProcessesResponse

Constructors

ResumeProcessesResponse