Network.AWS.AutoScalingPlans

Description

Derived from API version 2018-01-06 of the AWS service descriptions, licensed under Apache 2.0.

AWS Auto Scaling

Use AWS Auto Scaling to create scaling plans for your applications to automatically scale your scalable AWS resources.

API Summary

You can use the AWS Auto Scaling service API to accomplish the following tasks:

Create and manage scaling plans

Define target tracking scaling policies to dynamically scale your resources based on utilization

Scale Amazon EC2 Auto Scaling groups using predictive scaling and dynamic scaling to scale your Amazon EC2 capacity faster

Set minimum and maximum capacity limits

Retrieve information on existing scaling plans

Access current forecast data and historical forecast data for up to 56 days previous

To learn more about AWS Auto Scaling, including information about granting IAM users required permissions for AWS Auto Scaling actions, see the AWS Auto Scaling User Guide.

Service Configuration

defaultService :: Service

API version 2018-01-06 of the Amazon Auto Scaling Plans SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by AutoScalingPlans.

ObjectNotFoundException

_ObjectNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified object could not be found.

InternalServiceException

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The service encountered an internal error.

InvalidNextTokenException

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The token provided is not valid.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

An exception was thrown for a validation issue. Review the parameters provided.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Your account exceeded a limit. This exception is thrown when a per-account resource limit is exceeded.

ConcurrentUpdateException

_ConcurrentUpdateException :: AsError a => Getting (First ServiceError) a ServiceError

Concurrent updates caused an exception, for example, if you request an update to a scaling plan that already has a pending update.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

UpdateScalingPlan

data UpdateScalingPlan

See: newUpdateScalingPlan smart constructor.

Constructors

UpdateScalingPlan' (Maybe ApplicationSource) (Maybe [ScalingInstruction]) Text Integer

newUpdateScalingPlan

Arguments

:: Text
$sel:scalingPlanName:UpdateScalingPlan'

-> Integer
$sel:scalingPlanVersion:UpdateScalingPlan'

-> UpdateScalingPlan

Create a value of UpdateScalingPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationSource:UpdateScalingPlan', updateScalingPlan_applicationSource - A CloudFormation stack or set of tags.

For more information, see ApplicationSource in the AWS Auto Scaling API Reference.

$sel:scalingInstructions:UpdateScalingPlan', updateScalingPlan_scalingInstructions - The scaling instructions.

For more information, see ScalingInstruction in the AWS Auto Scaling API Reference.

$sel:scalingPlanName:UpdateScalingPlan', updateScalingPlan_scalingPlanName - The name of the scaling plan.

$sel:scalingPlanVersion:UpdateScalingPlan', updateScalingPlan_scalingPlanVersion - The version number of the scaling plan. The only valid value is 1. Currently, you cannot have multiple scaling plan versions.

data UpdateScalingPlanResponse

See: newUpdateScalingPlanResponse smart constructor.

Constructors

UpdateScalingPlanResponse' Int

newUpdateScalingPlanResponse

Arguments

:: Int
$sel:httpStatus:UpdateScalingPlanResponse'

-> UpdateScalingPlanResponse

Create a value of UpdateScalingPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateScalingPlanResponse', updateScalingPlanResponse_httpStatus - The response's http status code.

DeleteScalingPlan

data DeleteScalingPlan

See: newDeleteScalingPlan smart constructor.

Constructors

DeleteScalingPlan' Text Integer

newDeleteScalingPlan

Arguments

:: Text
$sel:scalingPlanName:DeleteScalingPlan'

-> Integer
$sel:scalingPlanVersion:DeleteScalingPlan'

-> DeleteScalingPlan

Create a value of DeleteScalingPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingPlanName:DeleteScalingPlan', deleteScalingPlan_scalingPlanName - The name of the scaling plan.

$sel:scalingPlanVersion:DeleteScalingPlan', deleteScalingPlan_scalingPlanVersion - The version number of the scaling plan. Currently, the only valid value is 1.

data DeleteScalingPlanResponse

See: newDeleteScalingPlanResponse smart constructor.

Constructors

DeleteScalingPlanResponse' Int

newDeleteScalingPlanResponse

Arguments

:: Int
$sel:httpStatus:DeleteScalingPlanResponse'

-> DeleteScalingPlanResponse

Create a value of DeleteScalingPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteScalingPlanResponse', deleteScalingPlanResponse_httpStatus - The response's http status code.

DescribeScalingPlanResources (Paginated)

data DescribeScalingPlanResources

See: newDescribeScalingPlanResources smart constructor.

Constructors

DescribeScalingPlanResources' (Maybe Text) (Maybe Int) Text Integer

newDescribeScalingPlanResources

Arguments

:: Text
$sel:scalingPlanName:DescribeScalingPlanResources'

-> Integer
$sel:scalingPlanVersion:DescribeScalingPlanResources'

-> DescribeScalingPlanResources

Create a value of DescribeScalingPlanResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeScalingPlanResources', describeScalingPlanResources_nextToken - The token for the next set of results.

$sel:maxResults:DescribeScalingPlanResources', describeScalingPlanResources_maxResults - The maximum number of scalable resources to return. The value must be between 1 and 50. The default value is 50.

$sel:scalingPlanName:DescribeScalingPlanResources', describeScalingPlanResources_scalingPlanName - The name of the scaling plan.

$sel:scalingPlanVersion:DescribeScalingPlanResources', describeScalingPlanResources_scalingPlanVersion - The version number of the scaling plan. Currently, the only valid value is 1.

data DescribeScalingPlanResourcesResponse

See: newDescribeScalingPlanResourcesResponse smart constructor.

Constructors

DescribeScalingPlanResourcesResponse' (Maybe Text) (Maybe [ScalingPlanResource]) Int

newDescribeScalingPlanResourcesResponse

Arguments

:: Int
$sel:httpStatus:DescribeScalingPlanResourcesResponse'

-> DescribeScalingPlanResourcesResponse

Create a value of DescribeScalingPlanResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeScalingPlanResources', describeScalingPlanResourcesResponse_nextToken - The token required to get the next set of results. This value is null if there are no more results to return.

$sel:scalingPlanResources:DescribeScalingPlanResourcesResponse', describeScalingPlanResourcesResponse_scalingPlanResources - Information about the scalable resources.

$sel:httpStatus:DescribeScalingPlanResourcesResponse', describeScalingPlanResourcesResponse_httpStatus - The response's http status code.

GetScalingPlanResourceForecastData

data GetScalingPlanResourceForecastData

See: newGetScalingPlanResourceForecastData smart constructor.

Constructors

GetScalingPlanResourceForecastData' Text Integer ServiceNamespace Text ScalableDimension ForecastDataType POSIX POSIX

newGetScalingPlanResourceForecastData

Arguments

:: Text
$sel:scalingPlanName:GetScalingPlanResourceForecastData'

-> Integer
$sel:scalingPlanVersion:GetScalingPlanResourceForecastData'

-> ServiceNamespace
$sel:serviceNamespace:GetScalingPlanResourceForecastData'

-> Text
$sel:resourceId:GetScalingPlanResourceForecastData'

-> ScalableDimension
$sel:scalableDimension:GetScalingPlanResourceForecastData'

-> ForecastDataType
$sel:forecastDataType:GetScalingPlanResourceForecastData'

-> UTCTime
$sel:startTime:GetScalingPlanResourceForecastData'

-> UTCTime
$sel:endTime:GetScalingPlanResourceForecastData'

-> GetScalingPlanResourceForecastData

Create a value of GetScalingPlanResourceForecastData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingPlanName:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_scalingPlanName - The name of the scaling plan.

$sel:scalingPlanVersion:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_scalingPlanVersion - The version number of the scaling plan. Currently, the only valid value is 1.

$sel:serviceNamespace:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_serviceNamespace - The namespace of the AWS service. The only valid value is autoscaling.

$sel:resourceId:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_resourceId - The ID of the resource. This string consists of a prefix (autoScalingGroup) followed by the name of a specified Auto Scaling group (my-asg). Example: autoScalingGroup/my-asg.

$sel:scalableDimension:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_scalableDimension - The scalable dimension for the resource. The only valid value is autoscaling:autoScalingGroup:DesiredCapacity.

$sel:forecastDataType:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_forecastDataType - The type of forecast data to get.

LoadForecast: The load metric forecast.

CapacityForecast: The capacity forecast.

ScheduledActionMinCapacity: The minimum capacity for each scheduled scaling action. This data is calculated as the larger of two values: the capacity forecast or the minimum capacity in the scaling instruction.

ScheduledActionMaxCapacity: The maximum capacity for each scheduled scaling action. The calculation used is determined by the predictive scaling maximum capacity behavior setting in the scaling instruction.

$sel:startTime:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_startTime - The inclusive start time of the time range for the forecast data to get. The date and time can be at most 56 days before the current date and time.

$sel:endTime:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_endTime - The exclusive end time of the time range for the forecast data to get. The maximum time duration between the start and end time is seven days.

Although this parameter can accept a date and time that is more than two days in the future, the availability of forecast data has limits. AWS Auto Scaling only issues forecasts for periods of two days in advance.

data GetScalingPlanResourceForecastDataResponse

See: newGetScalingPlanResourceForecastDataResponse smart constructor.

Constructors

GetScalingPlanResourceForecastDataResponse' Int [Datapoint]

newGetScalingPlanResourceForecastDataResponse

Arguments

:: Int
$sel:httpStatus:GetScalingPlanResourceForecastDataResponse'

-> GetScalingPlanResourceForecastDataResponse

Create a value of GetScalingPlanResourceForecastDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetScalingPlanResourceForecastDataResponse', getScalingPlanResourceForecastDataResponse_httpStatus - The response's http status code.

$sel:datapoints:GetScalingPlanResourceForecastDataResponse', getScalingPlanResourceForecastDataResponse_datapoints - The data points to return.

DescribeScalingPlans (Paginated)

data DescribeScalingPlans

See: newDescribeScalingPlans smart constructor.

Constructors

DescribeScalingPlans' (Maybe Integer) (Maybe Text) (Maybe Int) (Maybe [Text]) (Maybe [ApplicationSource])

newDescribeScalingPlans :: DescribeScalingPlans

Create a value of DescribeScalingPlans with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingPlanVersion:DescribeScalingPlans', describeScalingPlans_scalingPlanVersion - The version number of the scaling plan. Currently, the only valid value is 1.

If you specify a scaling plan version, you must also specify a scaling plan name.

$sel:nextToken:DescribeScalingPlans', describeScalingPlans_nextToken - The token for the next set of results.

$sel:maxResults:DescribeScalingPlans', describeScalingPlans_maxResults - The maximum number of scalable resources to return. This value can be between 1 and 50. The default value is 50.

$sel:scalingPlanNames:DescribeScalingPlans', describeScalingPlans_scalingPlanNames - The names of the scaling plans (up to 10). If you specify application sources, you cannot specify scaling plan names.

$sel:applicationSources:DescribeScalingPlans', describeScalingPlans_applicationSources - The sources for the applications (up to 10). If you specify scaling plan names, you cannot specify application sources.

data DescribeScalingPlansResponse

See: newDescribeScalingPlansResponse smart constructor.

Constructors

DescribeScalingPlansResponse' (Maybe Text) (Maybe [ScalingPlan]) Int

newDescribeScalingPlansResponse

Arguments

:: Int
$sel:httpStatus:DescribeScalingPlansResponse'

-> DescribeScalingPlansResponse

Create a value of DescribeScalingPlansResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeScalingPlans', describeScalingPlansResponse_nextToken - The token required to get the next set of results. This value is null if there are no more results to return.

$sel:scalingPlans:DescribeScalingPlansResponse', describeScalingPlansResponse_scalingPlans - Information about the scaling plans.

$sel:httpStatus:DescribeScalingPlansResponse', describeScalingPlansResponse_httpStatus - The response's http status code.

CreateScalingPlan

data CreateScalingPlan

See: newCreateScalingPlan smart constructor.

Constructors

CreateScalingPlan' Text ApplicationSource [ScalingInstruction]

newCreateScalingPlan

Arguments

:: Text
$sel:scalingPlanName:CreateScalingPlan'

-> ApplicationSource
$sel:applicationSource:CreateScalingPlan'

-> CreateScalingPlan

Create a value of CreateScalingPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingPlanName:CreateScalingPlan', createScalingPlan_scalingPlanName - The name of the scaling plan. Names cannot contain vertical bars, colons, or forward slashes.

$sel:applicationSource:CreateScalingPlan', createScalingPlan_applicationSource - A CloudFormation stack or set of tags. You can create one scaling plan per application source.

For more information, see ApplicationSource in the AWS Auto Scaling API Reference.

$sel:scalingInstructions:CreateScalingPlan', createScalingPlan_scalingInstructions - The scaling instructions.

For more information, see ScalingInstruction in the AWS Auto Scaling API Reference.

data CreateScalingPlanResponse

See: newCreateScalingPlanResponse smart constructor.

Constructors

CreateScalingPlanResponse' Int Integer

newCreateScalingPlanResponse

Arguments

:: Int
$sel:httpStatus:CreateScalingPlanResponse'

-> Integer
$sel:scalingPlanVersion:CreateScalingPlanResponse'

-> CreateScalingPlanResponse

Create a value of CreateScalingPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateScalingPlanResponse', createScalingPlanResponse_httpStatus - The response's http status code.

$sel:scalingPlanVersion:CreateScalingPlanResponse', createScalingPlanResponse_scalingPlanVersion - The version number of the scaling plan. This value is always 1. Currently, you cannot have multiple scaling plan versions.

Types

ForecastDataType

newtype ForecastDataType

Constructors

ForecastDataType'

Fields

fromForecastDataType :: Text

Bundled Patterns

pattern ForecastDataType_CapacityForecast :: ForecastDataType

pattern ForecastDataType_LoadForecast :: ForecastDataType

pattern ForecastDataType_ScheduledActionMaxCapacity :: ForecastDataType

pattern ForecastDataType_ScheduledActionMinCapacity :: ForecastDataType

LoadMetricType

newtype LoadMetricType

Constructors

LoadMetricType'

Fields

fromLoadMetricType :: Text

Bundled Patterns

pattern LoadMetricType_ALBTargetGroupRequestCount :: LoadMetricType

pattern LoadMetricType_ASGTotalCPUUtilization :: LoadMetricType

pattern LoadMetricType_ASGTotalNetworkIn :: LoadMetricType

pattern LoadMetricType_ASGTotalNetworkOut :: LoadMetricType

MetricStatistic

newtype MetricStatistic

Constructors

MetricStatistic'

Fields

fromMetricStatistic :: Text

Bundled Patterns

pattern MetricStatistic_Average :: MetricStatistic

pattern MetricStatistic_Maximum :: MetricStatistic

pattern MetricStatistic_Minimum :: MetricStatistic

pattern MetricStatistic_SampleCount :: MetricStatistic

pattern MetricStatistic_Sum :: MetricStatistic

PolicyType

newtype PolicyType

Constructors

PolicyType'

Fields

fromPolicyType :: Text

Bundled Patterns

pattern PolicyType_TargetTrackingScaling :: PolicyType

PredictiveScalingMaxCapacityBehavior

newtype PredictiveScalingMaxCapacityBehavior

Constructors

PredictiveScalingMaxCapacityBehavior'

Fields

fromPredictiveScalingMaxCapacityBehavior :: Text

Bundled Patterns

pattern PredictiveScalingMaxCapacityBehavior_SetForecastCapacityToMaxCapacity :: PredictiveScalingMaxCapacityBehavior

pattern PredictiveScalingMaxCapacityBehavior_SetMaxCapacityAboveForecastCapacity :: PredictiveScalingMaxCapacityBehavior

pattern PredictiveScalingMaxCapacityBehavior_SetMaxCapacityToForecastCapacity :: PredictiveScalingMaxCapacityBehavior

PredictiveScalingMode

newtype PredictiveScalingMode

Constructors

PredictiveScalingMode'

Fields

fromPredictiveScalingMode :: Text

Bundled Patterns

pattern PredictiveScalingMode_ForecastAndScale :: PredictiveScalingMode

pattern PredictiveScalingMode_ForecastOnly :: PredictiveScalingMode

ScalableDimension

newtype ScalableDimension

Constructors

ScalableDimension'

Fields

fromScalableDimension :: Text

Bundled Patterns

pattern ScalableDimension_Autoscaling_autoScalingGroup_DesiredCapacity :: ScalableDimension

pattern ScalableDimension_Dynamodb_index_ReadCapacityUnits :: ScalableDimension

pattern ScalableDimension_Dynamodb_index_WriteCapacityUnits :: ScalableDimension

pattern ScalableDimension_Dynamodb_table_ReadCapacityUnits :: ScalableDimension

pattern ScalableDimension_Dynamodb_table_WriteCapacityUnits :: ScalableDimension

pattern ScalableDimension_Ec2_spot_fleet_request_TargetCapacity :: ScalableDimension

pattern ScalableDimension_Ecs_service_DesiredCount :: ScalableDimension

pattern ScalableDimension_Rds_cluster_ReadReplicaCount :: ScalableDimension

ScalingMetricType

newtype ScalingMetricType

Constructors

ScalingMetricType'

Fields

fromScalingMetricType :: Text

Bundled Patterns

pattern ScalingMetricType_ALBRequestCountPerTarget :: ScalingMetricType

pattern ScalingMetricType_ASGAverageCPUUtilization :: ScalingMetricType

pattern ScalingMetricType_ASGAverageNetworkIn :: ScalingMetricType

pattern ScalingMetricType_ASGAverageNetworkOut :: ScalingMetricType

pattern ScalingMetricType_DynamoDBReadCapacityUtilization :: ScalingMetricType

pattern ScalingMetricType_DynamoDBWriteCapacityUtilization :: ScalingMetricType

pattern ScalingMetricType_EC2SpotFleetRequestAverageCPUUtilization :: ScalingMetricType

pattern ScalingMetricType_EC2SpotFleetRequestAverageNetworkIn :: ScalingMetricType

pattern ScalingMetricType_EC2SpotFleetRequestAverageNetworkOut :: ScalingMetricType

pattern ScalingMetricType_ECSServiceAverageCPUUtilization :: ScalingMetricType

pattern ScalingMetricType_ECSServiceAverageMemoryUtilization :: ScalingMetricType

pattern ScalingMetricType_RDSReaderAverageCPUUtilization :: ScalingMetricType

pattern ScalingMetricType_RDSReaderAverageDatabaseConnections :: ScalingMetricType

ScalingPlanStatusCode

newtype ScalingPlanStatusCode

Constructors

ScalingPlanStatusCode'

Fields

fromScalingPlanStatusCode :: Text

Bundled Patterns

pattern ScalingPlanStatusCode_Active :: ScalingPlanStatusCode

pattern ScalingPlanStatusCode_ActiveWithProblems :: ScalingPlanStatusCode

pattern ScalingPlanStatusCode_CreationFailed :: ScalingPlanStatusCode

pattern ScalingPlanStatusCode_CreationInProgress :: ScalingPlanStatusCode

pattern ScalingPlanStatusCode_DeletionFailed :: ScalingPlanStatusCode

pattern ScalingPlanStatusCode_DeletionInProgress :: ScalingPlanStatusCode

pattern ScalingPlanStatusCode_UpdateFailed :: ScalingPlanStatusCode

pattern ScalingPlanStatusCode_UpdateInProgress :: ScalingPlanStatusCode

ScalingPolicyUpdateBehavior

newtype ScalingPolicyUpdateBehavior

Constructors

ScalingPolicyUpdateBehavior'

Fields

fromScalingPolicyUpdateBehavior :: Text

Bundled Patterns

pattern ScalingPolicyUpdateBehavior_KeepExternalPolicies :: ScalingPolicyUpdateBehavior

pattern ScalingPolicyUpdateBehavior_ReplaceExternalPolicies :: ScalingPolicyUpdateBehavior

ScalingStatusCode

newtype ScalingStatusCode

Constructors

ScalingStatusCode'

Fields

fromScalingStatusCode :: Text

Bundled Patterns

pattern ScalingStatusCode_Active :: ScalingStatusCode

pattern ScalingStatusCode_Inactive :: ScalingStatusCode

pattern ScalingStatusCode_PartiallyActive :: ScalingStatusCode

ServiceNamespace

newtype ServiceNamespace

Constructors

ServiceNamespace'

Fields

fromServiceNamespace :: Text

Bundled Patterns

pattern ServiceNamespace_Autoscaling :: ServiceNamespace

pattern ServiceNamespace_Dynamodb :: ServiceNamespace

pattern ServiceNamespace_Ec2 :: ServiceNamespace

pattern ServiceNamespace_Ecs :: ServiceNamespace

pattern ServiceNamespace_Rds :: ServiceNamespace

ApplicationSource

data ApplicationSource

Represents an application source.

See: newApplicationSource smart constructor.

Constructors

ApplicationSource' (Maybe [TagFilter]) (Maybe Text)

newApplicationSource :: ApplicationSource

Create a value of ApplicationSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagFilters:ApplicationSource', applicationSource_tagFilters - A set of tags (up to 50).

$sel:cloudFormationStackARN:ApplicationSource', applicationSource_cloudFormationStackARN - The Amazon Resource Name (ARN) of a AWS CloudFormation stack.

CustomizedLoadMetricSpecification

data CustomizedLoadMetricSpecification

Represents a CloudWatch metric of your choosing that can be used for predictive scaling.

For predictive scaling to work with a customized load metric specification, AWS Auto Scaling needs access to the Sum and Average statistics that CloudWatch computes from metric data.

When you choose a load metric, make sure that the required Sum and Average statistics for your metric are available in CloudWatch and that they provide relevant data for predictive scaling. The Sum statistic must represent the total load on the resource, and the Average statistic must represent the average load per capacity unit of the resource. For example, there is a metric that counts the number of requests processed by your Auto Scaling group. If the Sum statistic represents the total request count processed by the group, then the Average statistic for the specified metric must represent the average request count processed by each instance of the group.

If you publish your own metrics, you can aggregate the data points at a given interval and then publish the aggregated data points to CloudWatch. Before AWS Auto Scaling generates the forecast, it sums up all the metric data points that occurred within each hour to match the granularity period that is used in the forecast (60 minutes).

For information about terminology, available metrics, or how to publish new metrics, see Amazon CloudWatch Concepts in the Amazon CloudWatch User Guide.

After creating your scaling plan, you can use the AWS Auto Scaling console to visualize forecasts for the specified metric. For more information, see View Scaling Information for a Resource in the AWS Auto Scaling User Guide.

See: newCustomizedLoadMetricSpecification smart constructor.

Constructors

CustomizedLoadMetricSpecification' (Maybe Text) (Maybe [MetricDimension]) Text Text MetricStatistic

newCustomizedLoadMetricSpecification

Arguments

:: Text
$sel:metricName:CustomizedLoadMetricSpecification'

-> Text
$sel:namespace:CustomizedLoadMetricSpecification'

-> MetricStatistic
$sel:statistic:CustomizedLoadMetricSpecification'

-> CustomizedLoadMetricSpecification

Create a value of CustomizedLoadMetricSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unit:CustomizedLoadMetricSpecification', customizedLoadMetricSpecification_unit - The unit of the metric.

$sel:dimensions:CustomizedLoadMetricSpecification', customizedLoadMetricSpecification_dimensions - The dimensions of the metric.

Conditional: If you published your metric with dimensions, you must specify the same dimensions in your customized load metric specification.

$sel:metricName:CustomizedLoadMetricSpecification', customizedLoadMetricSpecification_metricName - The name of the metric.

$sel:namespace:CustomizedLoadMetricSpecification', customizedLoadMetricSpecification_namespace - The namespace of the metric.

$sel:statistic:CustomizedLoadMetricSpecification', customizedLoadMetricSpecification_statistic - The statistic of the metric. The only valid value is Sum.

CustomizedScalingMetricSpecification

data CustomizedScalingMetricSpecification

Represents a CloudWatch metric of your choosing that can be used for dynamic scaling as part of a target tracking scaling policy.

To create your customized scaling metric specification:

Add values for each required parameter from CloudWatch. You can use an existing metric, or a new metric that you create. To use your own metric, you must first publish the metric to CloudWatch. For more information, see Publish Custom Metrics in the Amazon CloudWatch User Guide.

Choose a metric that changes proportionally with capacity. The value of the metric should increase or decrease in inverse proportion to the number of capacity units. That is, the value of the metric should decrease when capacity increases.

For information about terminology, available metrics, or how to publish new metrics, see Amazon CloudWatch Concepts in the Amazon CloudWatch User Guide.

See: newCustomizedScalingMetricSpecification smart constructor.

Constructors

CustomizedScalingMetricSpecification' (Maybe Text) (Maybe [MetricDimension]) Text Text MetricStatistic

newCustomizedScalingMetricSpecification

Arguments

:: Text
$sel:metricName:CustomizedScalingMetricSpecification'

-> Text
$sel:namespace:CustomizedScalingMetricSpecification'

-> MetricStatistic
$sel:statistic:CustomizedScalingMetricSpecification'

-> CustomizedScalingMetricSpecification

Create a value of CustomizedScalingMetricSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unit:CustomizedScalingMetricSpecification', customizedScalingMetricSpecification_unit - The unit of the metric.

$sel:dimensions:CustomizedScalingMetricSpecification', customizedScalingMetricSpecification_dimensions - The dimensions of the metric.

Conditional: If you published your metric with dimensions, you must specify the same dimensions in your customized scaling metric specification.

$sel:metricName:CustomizedScalingMetricSpecification', customizedScalingMetricSpecification_metricName - The name of the metric.

$sel:namespace:CustomizedScalingMetricSpecification', customizedScalingMetricSpecification_namespace - The namespace of the metric.

$sel:statistic:CustomizedScalingMetricSpecification', customizedScalingMetricSpecification_statistic - The statistic of the metric.

Datapoint

data Datapoint

Represents a single value in the forecast data used for predictive scaling.

See: newDatapoint smart constructor.

Constructors

Datapoint' (Maybe POSIX) (Maybe Double)

newDatapoint :: Datapoint

Create a value of Datapoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timestamp:Datapoint', datapoint_timestamp - The time stamp for the data point in UTC format.

$sel:value:Datapoint', datapoint_value - The value of the data point.

MetricDimension

data MetricDimension

Represents a dimension for a customized metric.

See: newMetricDimension smart constructor.

Constructors

MetricDimension' Text Text

newMetricDimension

Arguments

:: Text
$sel:name:MetricDimension'

-> Text
$sel:value:MetricDimension'

-> MetricDimension

Create a value of MetricDimension with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:MetricDimension', metricDimension_name - The name of the dimension.

$sel:value:MetricDimension', metricDimension_value - The value of the dimension.

PredefinedLoadMetricSpecification

data PredefinedLoadMetricSpecification

Represents a predefined metric that can be used for predictive scaling.

After creating your scaling plan, you can use the AWS Auto Scaling console to visualize forecasts for the specified metric. For more information, see View Scaling Information for a Resource in the AWS Auto Scaling User Guide.

See: newPredefinedLoadMetricSpecification smart constructor.

Constructors

PredefinedLoadMetricSpecification' (Maybe Text) LoadMetricType

newPredefinedLoadMetricSpecification

Arguments

:: LoadMetricType
$sel:predefinedLoadMetricType:PredefinedLoadMetricSpecification'

-> PredefinedLoadMetricSpecification

Create a value of PredefinedLoadMetricSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceLabel:PredefinedLoadMetricSpecification', predefinedLoadMetricSpecification_resourceLabel - Identifies the resource associated with the metric type. You can't specify a resource label unless the metric type is ALBTargetGroupRequestCount and there is a target group for an Application Load Balancer attached to the Auto Scaling group.

You create the resource label by appending the final portion of the load balancer ARN and the final portion of the target group ARN into a single value, separated by a forward slash (/). The format is app/<load-balancer-name>/<load-balancer-id>/targetgroup/<target-group-name>/<target-group-id>, where:

app/<load-balancer-name>/<load-balancer-id> is the final portion of the load balancer ARN

targetgroup/<target-group-name>/<target-group-id> is the final portion of the target group ARN.

This is an example: app/EC2Co-EcsEl-1TKLTMITMM0EO/f37c06a68c1748aa/targetgroup/EC2Co-Defau-LDNM7Q3ZH1ZN/6d4ea56ca2d6a18d.

To find the ARN for an Application Load Balancer, use the DescribeLoadBalancers API operation. To find the ARN for the target group, use the DescribeTargetGroups API operation.

$sel:predefinedLoadMetricType:PredefinedLoadMetricSpecification', predefinedLoadMetricSpecification_predefinedLoadMetricType - The metric type.

PredefinedScalingMetricSpecification

data PredefinedScalingMetricSpecification

Represents a predefined metric that can be used for dynamic scaling as part of a target tracking scaling policy.

See: newPredefinedScalingMetricSpecification smart constructor.

Constructors

PredefinedScalingMetricSpecification' (Maybe Text) ScalingMetricType

newPredefinedScalingMetricSpecification

Arguments

:: ScalingMetricType
$sel:predefinedScalingMetricType:PredefinedScalingMetricSpecification'

-> PredefinedScalingMetricSpecification

Create a value of PredefinedScalingMetricSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceLabel:PredefinedScalingMetricSpecification', predefinedScalingMetricSpecification_resourceLabel - Identifies the resource associated with the metric type. You can't specify a resource label unless the metric type is ALBRequestCountPerTarget and there is a target group for an Application Load Balancer attached to the Auto Scaling group, Spot Fleet request, or ECS service.

You create the resource label by appending the final portion of the load balancer ARN and the final portion of the target group ARN into a single value, separated by a forward slash (/). The format is app/<load-balancer-name>/<load-balancer-id>/targetgroup/<target-group-name>/<target-group-id>, where:

app/<load-balancer-name>/<load-balancer-id> is the final portion of the load balancer ARN

targetgroup/<target-group-name>/<target-group-id> is the final portion of the target group ARN.

This is an example: app/EC2Co-EcsEl-1TKLTMITMM0EO/f37c06a68c1748aa/targetgroup/EC2Co-Defau-LDNM7Q3ZH1ZN/6d4ea56ca2d6a18d.

To find the ARN for an Application Load Balancer, use the DescribeLoadBalancers API operation. To find the ARN for the target group, use the DescribeTargetGroups API operation.

$sel:predefinedScalingMetricType:PredefinedScalingMetricSpecification', predefinedScalingMetricSpecification_predefinedScalingMetricType - The metric type. The ALBRequestCountPerTarget metric type applies only to Auto Scaling groups, Spot Fleet requests, and ECS services.

ScalingInstruction

data ScalingInstruction

Describes a scaling instruction for a scalable resource in a scaling plan. Each scaling instruction applies to one resource.

AWS Auto Scaling creates target tracking scaling policies based on the scaling instructions. Target tracking scaling policies adjust the capacity of your scalable resource as required to maintain resource utilization at the target value that you specified.

AWS Auto Scaling also configures predictive scaling for your Amazon EC2 Auto Scaling groups using a subset of parameters, including the load metric, the scaling metric, the target value for the scaling metric, the predictive scaling mode (forecast and scale or forecast only), and the desired behavior when the forecast capacity exceeds the maximum capacity of the resource. With predictive scaling, AWS Auto Scaling generates forecasts with traffic predictions for the two days ahead and schedules scaling actions that proactively add and remove resource capacity to match the forecast.

We recommend waiting a minimum of 24 hours after creating an Auto Scaling group to configure predictive scaling. At minimum, there must be 24 hours of historical data to generate a forecast. For more information, see Best Practices for AWS Auto Scaling in the AWS Auto Scaling User Guide.

See: newScalingInstruction smart constructor.

Constructors

ScalingInstruction' (Maybe Bool) (Maybe PredefinedLoadMetricSpecification) (Maybe PredictiveScalingMaxCapacityBehavior) (Maybe CustomizedLoadMetricSpecification) (Maybe Int) (Maybe PredictiveScalingMode) (Maybe ScalingPolicyUpdateBehavior) (Maybe Natural) ServiceNamespace Text ScalableDimension Int Int [TargetTrackingConfiguration]

newScalingInstruction

Arguments

:: ServiceNamespace
$sel:serviceNamespace:ScalingInstruction'

-> Text
$sel:resourceId:ScalingInstruction'

-> ScalableDimension
$sel:scalableDimension:ScalingInstruction'

-> Int
$sel:minCapacity:ScalingInstruction'

-> Int
$sel:maxCapacity:ScalingInstruction'

-> ScalingInstruction

Create a value of ScalingInstruction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disableDynamicScaling:ScalingInstruction', scalingInstruction_disableDynamicScaling - Controls whether dynamic scaling by AWS Auto Scaling is disabled. When dynamic scaling is enabled, AWS Auto Scaling creates target tracking scaling policies based on the specified target tracking configurations.

The default is enabled (false).

$sel:predefinedLoadMetricSpecification:ScalingInstruction', scalingInstruction_predefinedLoadMetricSpecification - The predefined load metric to use for predictive scaling. This parameter or a CustomizedLoadMetricSpecification is required when configuring predictive scaling, and cannot be used otherwise.

$sel:predictiveScalingMaxCapacityBehavior:ScalingInstruction', scalingInstruction_predictiveScalingMaxCapacityBehavior - Defines the behavior that should be applied if the forecast capacity approaches or exceeds the maximum capacity specified for the resource. The default value is SetForecastCapacityToMaxCapacity.

The following are possible values:

SetForecastCapacityToMaxCapacity - AWS Auto Scaling cannot scale resource capacity higher than the maximum capacity. The maximum capacity is enforced as a hard limit.

SetMaxCapacityToForecastCapacity - AWS Auto Scaling may scale resource capacity higher than the maximum capacity to equal but not exceed forecast capacity.

SetMaxCapacityAboveForecastCapacity - AWS Auto Scaling may scale resource capacity higher than the maximum capacity by a specified buffer value. The intention is to give the target tracking scaling policy extra capacity if unexpected traffic occurs.

Only valid when configuring predictive scaling.

$sel:customizedLoadMetricSpecification:ScalingInstruction', scalingInstruction_customizedLoadMetricSpecification - The customized load metric to use for predictive scaling. This parameter or a PredefinedLoadMetricSpecification is required when configuring predictive scaling, and cannot be used otherwise.

$sel:predictiveScalingMaxCapacityBuffer:ScalingInstruction', scalingInstruction_predictiveScalingMaxCapacityBuffer - The size of the capacity buffer to use when the forecast capacity is close to or exceeds the maximum capacity. The value is specified as a percentage relative to the forecast capacity. For example, if the buffer is 10, this means a 10 percent buffer, such that if the forecast capacity is 50, and the maximum capacity is 40, then the effective maximum capacity is 55.

Only valid when configuring predictive scaling. Required if the PredictiveScalingMaxCapacityBehavior is set to SetMaxCapacityAboveForecastCapacity, and cannot be used otherwise.

The range is 1-100.

$sel:predictiveScalingMode:ScalingInstruction', scalingInstruction_predictiveScalingMode - The predictive scaling mode. The default value is ForecastAndScale. Otherwise, AWS Auto Scaling forecasts capacity but does not create any scheduled scaling actions based on the capacity forecast.

$sel:scalingPolicyUpdateBehavior:ScalingInstruction', scalingInstruction_scalingPolicyUpdateBehavior - Controls whether a resource's externally created scaling policies are kept or replaced.

The default value is KeepExternalPolicies. If the parameter is set to ReplaceExternalPolicies, any scaling policies that are external to AWS Auto Scaling are deleted and new target tracking scaling policies created.

Only valid when configuring dynamic scaling.

Condition: The number of existing policies to be replaced must be less than or equal to 50. If there are more than 50 policies to be replaced, AWS Auto Scaling keeps all existing policies and does not create new ones.

$sel:scheduledActionBufferTime:ScalingInstruction', scalingInstruction_scheduledActionBufferTime - The amount of time, in seconds, to buffer the run time of scheduled scaling actions when scaling out. For example, if the forecast says to add capacity at 10:00 AM, and the buffer time is 5 minutes, then the run time of the corresponding scheduled scaling action will be 9:55 AM. The intention is to give resources time to be provisioned. For example, it can take a few minutes to launch an EC2 instance. The actual amount of time required depends on several factors, such as the size of the instance and whether there are startup scripts to complete.

The value must be less than the forecast interval duration of 3600 seconds (60 minutes). The default is 300 seconds.

Only valid when configuring predictive scaling.

$sel:serviceNamespace:ScalingInstruction', scalingInstruction_serviceNamespace - The namespace of the AWS service.

$sel:resourceId:ScalingInstruction', scalingInstruction_resourceId - The ID of the resource. This string consists of the resource type and unique identifier.

Auto Scaling group - The resource type is autoScalingGroup and the unique identifier is the name of the Auto Scaling group. Example: autoScalingGroup/my-asg.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

$sel:scalableDimension:ScalingInstruction', scalingInstruction_scalableDimension - The scalable dimension associated with the resource.

autoscaling:autoScalingGroup:DesiredCapacity - The desired capacity of an Auto Scaling group.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet request.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

$sel:minCapacity:ScalingInstruction', scalingInstruction_minCapacity - The minimum capacity of the resource.

$sel:maxCapacity:ScalingInstruction', scalingInstruction_maxCapacity - The maximum capacity of the resource. The exception to this upper limit is if you specify a non-default setting for PredictiveScalingMaxCapacityBehavior.

$sel:targetTrackingConfigurations:ScalingInstruction', scalingInstruction_targetTrackingConfigurations - The target tracking configurations (up to 10). Each of these structures must specify a unique scaling metric and a target value for the metric.

ScalingPlan

data ScalingPlan

Represents a scaling plan.

See: newScalingPlan smart constructor.

Constructors

ScalingPlan' (Maybe Text) (Maybe POSIX) (Maybe POSIX) Text Integer ApplicationSource [ScalingInstruction] ScalingPlanStatusCode

newScalingPlan

Arguments

:: Text
$sel:scalingPlanName:ScalingPlan'

-> Integer
$sel:scalingPlanVersion:ScalingPlan'

-> ApplicationSource
$sel:applicationSource:ScalingPlan'

-> ScalingPlanStatusCode
$sel:statusCode:ScalingPlan'

-> ScalingPlan

Create a value of ScalingPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:ScalingPlan', scalingPlan_statusMessage - A simple message about the current status of the scaling plan.

$sel:creationTime:ScalingPlan', scalingPlan_creationTime - The Unix time stamp when the scaling plan was created.

$sel:statusStartTime:ScalingPlan', scalingPlan_statusStartTime - The Unix time stamp when the scaling plan entered the current status.

$sel:scalingPlanName:ScalingPlan', scalingPlan_scalingPlanName - The name of the scaling plan.

$sel:scalingPlanVersion:ScalingPlan', scalingPlan_scalingPlanVersion - The version number of the scaling plan.

$sel:applicationSource:ScalingPlan', scalingPlan_applicationSource - A CloudFormation stack or a set of tags. You can create one scaling plan per application source.

$sel:scalingInstructions:ScalingPlan', scalingPlan_scalingInstructions - The scaling instructions.

$sel:statusCode:ScalingPlan', scalingPlan_statusCode - The status of the scaling plan.

Active - The scaling plan is active.

ActiveWithProblems - The scaling plan is active, but the scaling configuration for one or more resources could not be applied.

CreationInProgress - The scaling plan is being created.

CreationFailed - The scaling plan could not be created.

DeletionInProgress - The scaling plan is being deleted.

DeletionFailed - The scaling plan could not be deleted.

UpdateInProgress - The scaling plan is being updated.

UpdateFailed - The scaling plan could not be updated.

ScalingPlanResource

data ScalingPlanResource

Represents a scalable resource.

See: newScalingPlanResource smart constructor.

Constructors

ScalingPlanResource' (Maybe [ScalingPolicy]) (Maybe Text) Text Integer ServiceNamespace Text ScalableDimension ScalingStatusCode

newScalingPlanResource

Arguments

:: Text
$sel:scalingPlanName:ScalingPlanResource'

-> Integer
$sel:scalingPlanVersion:ScalingPlanResource'

-> ServiceNamespace
$sel:serviceNamespace:ScalingPlanResource'

-> Text
$sel:resourceId:ScalingPlanResource'

-> ScalableDimension
$sel:scalableDimension:ScalingPlanResource'

-> ScalingStatusCode
$sel:scalingStatusCode:ScalingPlanResource'

-> ScalingPlanResource

Create a value of ScalingPlanResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingPolicies:ScalingPlanResource', scalingPlanResource_scalingPolicies - The scaling policies.

$sel:scalingStatusMessage:ScalingPlanResource', scalingPlanResource_scalingStatusMessage - A simple message about the current scaling status of the resource.

$sel:scalingPlanName:ScalingPlanResource', scalingPlanResource_scalingPlanName - The name of the scaling plan.

$sel:scalingPlanVersion:ScalingPlanResource', scalingPlanResource_scalingPlanVersion - The version number of the scaling plan.

$sel:serviceNamespace:ScalingPlanResource', scalingPlanResource_serviceNamespace - The namespace of the AWS service.

$sel:resourceId:ScalingPlanResource', scalingPlanResource_resourceId - The ID of the resource. This string consists of the resource type and unique identifier.

Auto Scaling group - The resource type is autoScalingGroup and the unique identifier is the name of the Auto Scaling group. Example: autoScalingGroup/my-asg.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

$sel:scalableDimension:ScalingPlanResource', scalingPlanResource_scalableDimension - The scalable dimension for the resource.

autoscaling:autoScalingGroup:DesiredCapacity - The desired capacity of an Auto Scaling group.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet request.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

$sel:scalingStatusCode:ScalingPlanResource', scalingPlanResource_scalingStatusCode - The scaling status of the resource.

Active - The scaling configuration is active.

Inactive - The scaling configuration is not active because the scaling plan is being created or the scaling configuration could not be applied. Check the status message for more information.

PartiallyActive - The scaling configuration is partially active because the scaling plan is being created or deleted or the scaling configuration could not be fully applied. Check the status message for more information.

ScalingPolicy

data ScalingPolicy

Represents a scaling policy.

See: newScalingPolicy smart constructor.

Constructors

ScalingPolicy' (Maybe TargetTrackingConfiguration) Text PolicyType

newScalingPolicy

Arguments

:: Text
$sel:policyName:ScalingPolicy'

-> PolicyType
$sel:policyType:ScalingPolicy'

-> ScalingPolicy

Create a value of ScalingPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetTrackingConfiguration:ScalingPolicy', scalingPolicy_targetTrackingConfiguration - The target tracking scaling policy. Includes support for predefined or customized metrics.

$sel:policyName:ScalingPolicy', scalingPolicy_policyName - The name of the scaling policy.

$sel:policyType:ScalingPolicy', scalingPolicy_policyType - The type of scaling policy.

TagFilter

data TagFilter

Represents a tag.

See: newTagFilter smart constructor.

Constructors

TagFilter' (Maybe Text) (Maybe [Text])

newTagFilter :: TagFilter

Create a value of TagFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:TagFilter', tagFilter_key - The tag key.

$sel:values:TagFilter', tagFilter_values - The tag values (0 to 20).

TargetTrackingConfiguration

data TargetTrackingConfiguration

Describes a target tracking configuration to use with AWS Auto Scaling. Used with ScalingInstruction and ScalingPolicy.

See: newTargetTrackingConfiguration smart constructor.

Constructors

TargetTrackingConfiguration' (Maybe Bool) (Maybe CustomizedScalingMetricSpecification) (Maybe PredefinedScalingMetricSpecification) (Maybe Int) (Maybe Int) (Maybe Int) Double

newTargetTrackingConfiguration

Arguments

:: Double
$sel:targetValue:TargetTrackingConfiguration'

-> TargetTrackingConfiguration

Create a value of TargetTrackingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disableScaleIn:TargetTrackingConfiguration', targetTrackingConfiguration_disableScaleIn - Indicates whether scale in by the target tracking scaling policy is disabled. If the value is true, scale in is disabled and the target tracking scaling policy doesn't remove capacity from the scalable resource. Otherwise, scale in is enabled and the target tracking scaling policy can remove capacity from the scalable resource.

The default value is false.

$sel:customizedScalingMetricSpecification:TargetTrackingConfiguration', targetTrackingConfiguration_customizedScalingMetricSpecification - A customized metric. You can specify either a predefined metric or a customized metric.

$sel:predefinedScalingMetricSpecification:TargetTrackingConfiguration', targetTrackingConfiguration_predefinedScalingMetricSpecification - A predefined metric. You can specify either a predefined metric or a customized metric.

$sel:estimatedInstanceWarmup:TargetTrackingConfiguration', targetTrackingConfiguration_estimatedInstanceWarmup - The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics. This value is used only if the resource is an Auto Scaling group.

$sel:scaleOutCooldown:TargetTrackingConfiguration', targetTrackingConfiguration_scaleOutCooldown - The amount of time, in seconds, to wait for a previous scale-out activity to take effect. This property is not used if the scalable resource is an Auto Scaling group.

With the scale-out cooldown period, the intention is to continuously (but not excessively) scale out. After Auto Scaling successfully scales out using a target tracking scaling policy, it starts to calculate the cooldown time. The scaling policy won't increase the desired capacity again unless either a larger scale out is triggered or the cooldown period ends.

$sel:scaleInCooldown:TargetTrackingConfiguration', targetTrackingConfiguration_scaleInCooldown - The amount of time, in seconds, after a scale-in activity completes before another scale-in activity can start. This property is not used if the scalable resource is an Auto Scaling group.

With the scale-in cooldown period, the intention is to scale in conservatively to protect your application’s availability, so scale-in activities are blocked until the cooldown period has expired. However, if another alarm triggers a scale-out activity during the scale-in cooldown period, Auto Scaling scales out the target immediately. In this case, the scale-in cooldown period stops and doesn't complete.

$sel:targetValue:TargetTrackingConfiguration', targetTrackingConfiguration_targetValue - The target value for the metric. Although this property accepts numbers of type Double, it won't accept values that are either too small or too large. Values must be in the range of -2^360 to 2^360.