Network.AWS.AutoScalingPlans.DescribeScalingPlans

Description

Describes one or more of your scaling plans.

This operation returns paginated results.

Creating a Request

data DescribeScalingPlans

See: newDescribeScalingPlans smart constructor.

Constructors

DescribeScalingPlans'

Fields

scalingPlanVersion :: Maybe Integer

The version number of the scaling plan. Currently, the only valid value is 1.

If you specify a scaling plan version, you must also specify a scaling plan name.

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Int

The maximum number of scalable resources to return. This value can be between 1 and 50. The default value is 50.

scalingPlanNames :: Maybe [Text]

The names of the scaling plans (up to 10). If you specify application sources, you cannot specify scaling plan names.

applicationSources :: Maybe [ApplicationSource]

The sources for the applications (up to 10). If you specify scaling plan names, you cannot specify application sources.

newDescribeScalingPlans :: DescribeScalingPlans

Create a value of DescribeScalingPlans with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingPlanVersion:DescribeScalingPlans', describeScalingPlans_scalingPlanVersion - The version number of the scaling plan. Currently, the only valid value is 1.

If you specify a scaling plan version, you must also specify a scaling plan name.

$sel:nextToken:DescribeScalingPlans', describeScalingPlans_nextToken - The token for the next set of results.

$sel:maxResults:DescribeScalingPlans', describeScalingPlans_maxResults - The maximum number of scalable resources to return. This value can be between 1 and 50. The default value is 50.

$sel:scalingPlanNames:DescribeScalingPlans', describeScalingPlans_scalingPlanNames - The names of the scaling plans (up to 10). If you specify application sources, you cannot specify scaling plan names.

$sel:applicationSources:DescribeScalingPlans', describeScalingPlans_applicationSources - The sources for the applications (up to 10). If you specify scaling plan names, you cannot specify application sources.

Request Lenses

describeScalingPlans_scalingPlanVersion :: Lens' DescribeScalingPlans (Maybe Integer)

The version number of the scaling plan. Currently, the only valid value is 1.

If you specify a scaling plan version, you must also specify a scaling plan name.

describeScalingPlans_nextToken :: Lens' DescribeScalingPlans (Maybe Text)

The token for the next set of results.

describeScalingPlans_maxResults :: Lens' DescribeScalingPlans (Maybe Int)

The maximum number of scalable resources to return. This value can be between 1 and 50. The default value is 50.

describeScalingPlans_scalingPlanNames :: Lens' DescribeScalingPlans (Maybe [Text])

The names of the scaling plans (up to 10). If you specify application sources, you cannot specify scaling plan names.

describeScalingPlans_applicationSources :: Lens' DescribeScalingPlans (Maybe [ApplicationSource])

The sources for the applications (up to 10). If you specify scaling plan names, you cannot specify application sources.

Destructuring the Response

data DescribeScalingPlansResponse

See: newDescribeScalingPlansResponse smart constructor.

Constructors

DescribeScalingPlansResponse'

Fields

nextToken :: Maybe Text

The token required to get the next set of results. This value is null if there are no more results to return.

scalingPlans :: Maybe [ScalingPlan]

Information about the scaling plans.

httpStatus :: Int

The response's http status code.

newDescribeScalingPlansResponse

Arguments

:: Int
$sel:httpStatus:DescribeScalingPlansResponse'

-> DescribeScalingPlansResponse

Create a value of DescribeScalingPlansResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeScalingPlans', describeScalingPlansResponse_nextToken - The token required to get the next set of results. This value is null if there are no more results to return.

$sel:scalingPlans:DescribeScalingPlansResponse', describeScalingPlansResponse_scalingPlans - Information about the scaling plans.

$sel:httpStatus:DescribeScalingPlansResponse', describeScalingPlansResponse_httpStatus - The response's http status code.

Response Lenses

describeScalingPlansResponse_nextToken :: Lens' DescribeScalingPlansResponse (Maybe Text)

The token required to get the next set of results. This value is null if there are no more results to return.

describeScalingPlansResponse_scalingPlans :: Lens' DescribeScalingPlansResponse (Maybe [ScalingPlan])

Information about the scaling plans.

describeScalingPlansResponse_httpStatus :: Lens' DescribeScalingPlansResponse Int

The response's http status code.