Network.AWS.AutoScalingPlans.GetScalingPlanResourceForecastData

Description

Retrieves the forecast data for a scalable resource.

Capacity forecasts are represented as predicted values, or data points, that are calculated using historical data points from a specified CloudWatch load metric. Data points are available for up to 56 days.

Creating a Request

data GetScalingPlanResourceForecastData

See: newGetScalingPlanResourceForecastData smart constructor.

Constructors

GetScalingPlanResourceForecastData'

Fields

scalingPlanName :: Text

The name of the scaling plan.

scalingPlanVersion :: Integer

The version number of the scaling plan. Currently, the only valid value is 1.

serviceNamespace :: ServiceNamespace

The namespace of the AWS service. The only valid value is autoscaling.

resourceId :: Text

The ID of the resource. This string consists of a prefix (autoScalingGroup) followed by the name of a specified Auto Scaling group (my-asg). Example: autoScalingGroup/my-asg.

scalableDimension :: ScalableDimension

The scalable dimension for the resource. The only valid value is autoscaling:autoScalingGroup:DesiredCapacity.

forecastDataType :: ForecastDataType

The type of forecast data to get.

LoadForecast: The load metric forecast.

CapacityForecast: The capacity forecast.

ScheduledActionMinCapacity: The minimum capacity for each scheduled scaling action. This data is calculated as the larger of two values: the capacity forecast or the minimum capacity in the scaling instruction.

ScheduledActionMaxCapacity: The maximum capacity for each scheduled scaling action. The calculation used is determined by the predictive scaling maximum capacity behavior setting in the scaling instruction.

startTime :: POSIX

The inclusive start time of the time range for the forecast data to get. The date and time can be at most 56 days before the current date and time.

endTime :: POSIX

The exclusive end time of the time range for the forecast data to get. The maximum time duration between the start and end time is seven days.

Although this parameter can accept a date and time that is more than two days in the future, the availability of forecast data has limits. AWS Auto Scaling only issues forecasts for periods of two days in advance.

newGetScalingPlanResourceForecastData

Arguments

:: Text
$sel:scalingPlanName:GetScalingPlanResourceForecastData'

-> Integer
$sel:scalingPlanVersion:GetScalingPlanResourceForecastData'

-> ServiceNamespace
$sel:serviceNamespace:GetScalingPlanResourceForecastData'

-> Text
$sel:resourceId:GetScalingPlanResourceForecastData'

-> ScalableDimension
$sel:scalableDimension:GetScalingPlanResourceForecastData'

-> ForecastDataType
$sel:forecastDataType:GetScalingPlanResourceForecastData'

-> UTCTime
$sel:startTime:GetScalingPlanResourceForecastData'

-> UTCTime
$sel:endTime:GetScalingPlanResourceForecastData'

-> GetScalingPlanResourceForecastData

Create a value of GetScalingPlanResourceForecastData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingPlanName:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_scalingPlanName - The name of the scaling plan.

$sel:scalingPlanVersion:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_scalingPlanVersion - The version number of the scaling plan. Currently, the only valid value is 1.

$sel:serviceNamespace:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_serviceNamespace - The namespace of the AWS service. The only valid value is autoscaling.

$sel:resourceId:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_resourceId - The ID of the resource. This string consists of a prefix (autoScalingGroup) followed by the name of a specified Auto Scaling group (my-asg). Example: autoScalingGroup/my-asg.

$sel:scalableDimension:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_scalableDimension - The scalable dimension for the resource. The only valid value is autoscaling:autoScalingGroup:DesiredCapacity.

$sel:forecastDataType:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_forecastDataType - The type of forecast data to get.

LoadForecast: The load metric forecast.

CapacityForecast: The capacity forecast.

ScheduledActionMinCapacity: The minimum capacity for each scheduled scaling action. This data is calculated as the larger of two values: the capacity forecast or the minimum capacity in the scaling instruction.

ScheduledActionMaxCapacity: The maximum capacity for each scheduled scaling action. The calculation used is determined by the predictive scaling maximum capacity behavior setting in the scaling instruction.

$sel:startTime:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_startTime - The inclusive start time of the time range for the forecast data to get. The date and time can be at most 56 days before the current date and time.

$sel:endTime:GetScalingPlanResourceForecastData', getScalingPlanResourceForecastData_endTime - The exclusive end time of the time range for the forecast data to get. The maximum time duration between the start and end time is seven days.

Although this parameter can accept a date and time that is more than two days in the future, the availability of forecast data has limits. AWS Auto Scaling only issues forecasts for periods of two days in advance.

Request Lenses

getScalingPlanResourceForecastData_scalingPlanName :: Lens' GetScalingPlanResourceForecastData Text

The name of the scaling plan.

getScalingPlanResourceForecastData_scalingPlanVersion :: Lens' GetScalingPlanResourceForecastData Integer

The version number of the scaling plan. Currently, the only valid value is 1.

getScalingPlanResourceForecastData_serviceNamespace :: Lens' GetScalingPlanResourceForecastData ServiceNamespace

The namespace of the AWS service. The only valid value is autoscaling.

getScalingPlanResourceForecastData_resourceId :: Lens' GetScalingPlanResourceForecastData Text

The ID of the resource. This string consists of a prefix (autoScalingGroup) followed by the name of a specified Auto Scaling group (my-asg). Example: autoScalingGroup/my-asg.

getScalingPlanResourceForecastData_scalableDimension :: Lens' GetScalingPlanResourceForecastData ScalableDimension

The scalable dimension for the resource. The only valid value is autoscaling:autoScalingGroup:DesiredCapacity.

getScalingPlanResourceForecastData_forecastDataType :: Lens' GetScalingPlanResourceForecastData ForecastDataType

The type of forecast data to get.

LoadForecast: The load metric forecast.

CapacityForecast: The capacity forecast.

ScheduledActionMinCapacity: The minimum capacity for each scheduled scaling action. This data is calculated as the larger of two values: the capacity forecast or the minimum capacity in the scaling instruction.

ScheduledActionMaxCapacity: The maximum capacity for each scheduled scaling action. The calculation used is determined by the predictive scaling maximum capacity behavior setting in the scaling instruction.

getScalingPlanResourceForecastData_startTime :: Lens' GetScalingPlanResourceForecastData UTCTime

The inclusive start time of the time range for the forecast data to get. The date and time can be at most 56 days before the current date and time.

getScalingPlanResourceForecastData_endTime :: Lens' GetScalingPlanResourceForecastData UTCTime

The exclusive end time of the time range for the forecast data to get. The maximum time duration between the start and end time is seven days.

Although this parameter can accept a date and time that is more than two days in the future, the availability of forecast data has limits. AWS Auto Scaling only issues forecasts for periods of two days in advance.

Destructuring the Response

data GetScalingPlanResourceForecastDataResponse

See: newGetScalingPlanResourceForecastDataResponse smart constructor.

Constructors

GetScalingPlanResourceForecastDataResponse'

Fields

httpStatus :: Int

The response's http status code.

datapoints :: [Datapoint]

The data points to return.

newGetScalingPlanResourceForecastDataResponse

Arguments

:: Int
$sel:httpStatus:GetScalingPlanResourceForecastDataResponse'

-> GetScalingPlanResourceForecastDataResponse

Create a value of GetScalingPlanResourceForecastDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetScalingPlanResourceForecastDataResponse', getScalingPlanResourceForecastDataResponse_httpStatus - The response's http status code.

$sel:datapoints:GetScalingPlanResourceForecastDataResponse', getScalingPlanResourceForecastDataResponse_datapoints - The data points to return.

Response Lenses

getScalingPlanResourceForecastDataResponse_httpStatus :: Lens' GetScalingPlanResourceForecastDataResponse Int

The response's http status code.

getScalingPlanResourceForecastDataResponse_datapoints :: Lens' GetScalingPlanResourceForecastDataResponse [Datapoint]

The data points to return.