Network.AWS.AutoScalingPlans.Lens

Description

Operations

UpdateScalingPlan

updateScalingPlan_applicationSource :: Lens' UpdateScalingPlan (Maybe ApplicationSource)

A CloudFormation stack or set of tags.

For more information, see ApplicationSource in the AWS Auto Scaling API Reference.

updateScalingPlan_scalingInstructions :: Lens' UpdateScalingPlan (Maybe [ScalingInstruction])

The scaling instructions.

For more information, see ScalingInstruction in the AWS Auto Scaling API Reference.

updateScalingPlan_scalingPlanName :: Lens' UpdateScalingPlan Text

The name of the scaling plan.

updateScalingPlan_scalingPlanVersion :: Lens' UpdateScalingPlan Integer

The version number of the scaling plan. The only valid value is 1. Currently, you cannot have multiple scaling plan versions.

updateScalingPlanResponse_httpStatus :: Lens' UpdateScalingPlanResponse Int

The response's http status code.

DeleteScalingPlan

deleteScalingPlan_scalingPlanName :: Lens' DeleteScalingPlan Text

The name of the scaling plan.

deleteScalingPlan_scalingPlanVersion :: Lens' DeleteScalingPlan Integer

The version number of the scaling plan. Currently, the only valid value is 1.

deleteScalingPlanResponse_httpStatus :: Lens' DeleteScalingPlanResponse Int

The response's http status code.

DescribeScalingPlanResources

describeScalingPlanResources_nextToken :: Lens' DescribeScalingPlanResources (Maybe Text)

The token for the next set of results.

describeScalingPlanResources_maxResults :: Lens' DescribeScalingPlanResources (Maybe Int)

The maximum number of scalable resources to return. The value must be between 1 and 50. The default value is 50.

describeScalingPlanResources_scalingPlanName :: Lens' DescribeScalingPlanResources Text

The name of the scaling plan.

describeScalingPlanResources_scalingPlanVersion :: Lens' DescribeScalingPlanResources Integer

The version number of the scaling plan. Currently, the only valid value is 1.

describeScalingPlanResourcesResponse_nextToken :: Lens' DescribeScalingPlanResourcesResponse (Maybe Text)

The token required to get the next set of results. This value is null if there are no more results to return.

describeScalingPlanResourcesResponse_scalingPlanResources :: Lens' DescribeScalingPlanResourcesResponse (Maybe [ScalingPlanResource])

Information about the scalable resources.

describeScalingPlanResourcesResponse_httpStatus :: Lens' DescribeScalingPlanResourcesResponse Int

The response's http status code.

GetScalingPlanResourceForecastData

getScalingPlanResourceForecastData_scalingPlanName :: Lens' GetScalingPlanResourceForecastData Text

The name of the scaling plan.

getScalingPlanResourceForecastData_scalingPlanVersion :: Lens' GetScalingPlanResourceForecastData Integer

The version number of the scaling plan. Currently, the only valid value is 1.

getScalingPlanResourceForecastData_serviceNamespace :: Lens' GetScalingPlanResourceForecastData ServiceNamespace

The namespace of the AWS service. The only valid value is autoscaling.

getScalingPlanResourceForecastData_resourceId :: Lens' GetScalingPlanResourceForecastData Text

The ID of the resource. This string consists of a prefix (autoScalingGroup) followed by the name of a specified Auto Scaling group (my-asg). Example: autoScalingGroup/my-asg.

getScalingPlanResourceForecastData_scalableDimension :: Lens' GetScalingPlanResourceForecastData ScalableDimension

The scalable dimension for the resource. The only valid value is autoscaling:autoScalingGroup:DesiredCapacity.

getScalingPlanResourceForecastData_forecastDataType :: Lens' GetScalingPlanResourceForecastData ForecastDataType

The type of forecast data to get.

LoadForecast: The load metric forecast.

CapacityForecast: The capacity forecast.

ScheduledActionMinCapacity: The minimum capacity for each scheduled scaling action. This data is calculated as the larger of two values: the capacity forecast or the minimum capacity in the scaling instruction.

ScheduledActionMaxCapacity: The maximum capacity for each scheduled scaling action. The calculation used is determined by the predictive scaling maximum capacity behavior setting in the scaling instruction.

getScalingPlanResourceForecastData_startTime :: Lens' GetScalingPlanResourceForecastData UTCTime

The inclusive start time of the time range for the forecast data to get. The date and time can be at most 56 days before the current date and time.

getScalingPlanResourceForecastData_endTime :: Lens' GetScalingPlanResourceForecastData UTCTime

The exclusive end time of the time range for the forecast data to get. The maximum time duration between the start and end time is seven days.

Although this parameter can accept a date and time that is more than two days in the future, the availability of forecast data has limits. AWS Auto Scaling only issues forecasts for periods of two days in advance.

getScalingPlanResourceForecastDataResponse_httpStatus :: Lens' GetScalingPlanResourceForecastDataResponse Int

The response's http status code.

getScalingPlanResourceForecastDataResponse_datapoints :: Lens' GetScalingPlanResourceForecastDataResponse [Datapoint]

The data points to return.

DescribeScalingPlans

describeScalingPlans_scalingPlanVersion :: Lens' DescribeScalingPlans (Maybe Integer)

The version number of the scaling plan. Currently, the only valid value is 1.

If you specify a scaling plan version, you must also specify a scaling plan name.

describeScalingPlans_nextToken :: Lens' DescribeScalingPlans (Maybe Text)

The token for the next set of results.

describeScalingPlans_maxResults :: Lens' DescribeScalingPlans (Maybe Int)

The maximum number of scalable resources to return. This value can be between 1 and 50. The default value is 50.

describeScalingPlans_scalingPlanNames :: Lens' DescribeScalingPlans (Maybe [Text])

The names of the scaling plans (up to 10). If you specify application sources, you cannot specify scaling plan names.

describeScalingPlans_applicationSources :: Lens' DescribeScalingPlans (Maybe [ApplicationSource])

The sources for the applications (up to 10). If you specify scaling plan names, you cannot specify application sources.

describeScalingPlansResponse_nextToken :: Lens' DescribeScalingPlansResponse (Maybe Text)

The token required to get the next set of results. This value is null if there are no more results to return.

describeScalingPlansResponse_scalingPlans :: Lens' DescribeScalingPlansResponse (Maybe [ScalingPlan])

Information about the scaling plans.

describeScalingPlansResponse_httpStatus :: Lens' DescribeScalingPlansResponse Int

The response's http status code.

CreateScalingPlan

createScalingPlan_scalingPlanName :: Lens' CreateScalingPlan Text

The name of the scaling plan. Names cannot contain vertical bars, colons, or forward slashes.

createScalingPlan_applicationSource :: Lens' CreateScalingPlan ApplicationSource

A CloudFormation stack or set of tags. You can create one scaling plan per application source.

For more information, see ApplicationSource in the AWS Auto Scaling API Reference.

createScalingPlan_scalingInstructions :: Lens' CreateScalingPlan [ScalingInstruction]

The scaling instructions.

For more information, see ScalingInstruction in the AWS Auto Scaling API Reference.

createScalingPlanResponse_httpStatus :: Lens' CreateScalingPlanResponse Int

The response's http status code.

createScalingPlanResponse_scalingPlanVersion :: Lens' CreateScalingPlanResponse Integer

The version number of the scaling plan. This value is always 1. Currently, you cannot have multiple scaling plan versions.

Types

ApplicationSource

applicationSource_tagFilters :: Lens' ApplicationSource (Maybe [TagFilter])

A set of tags (up to 50).

applicationSource_cloudFormationStackARN :: Lens' ApplicationSource (Maybe Text)

The Amazon Resource Name (ARN) of a AWS CloudFormation stack.

CustomizedLoadMetricSpecification

customizedLoadMetricSpecification_unit :: Lens' CustomizedLoadMetricSpecification (Maybe Text)

The unit of the metric.

customizedLoadMetricSpecification_dimensions :: Lens' CustomizedLoadMetricSpecification (Maybe [MetricDimension])

The dimensions of the metric.

Conditional: If you published your metric with dimensions, you must specify the same dimensions in your customized load metric specification.

customizedLoadMetricSpecification_metricName :: Lens' CustomizedLoadMetricSpecification Text

The name of the metric.

customizedLoadMetricSpecification_namespace :: Lens' CustomizedLoadMetricSpecification Text

The namespace of the metric.

customizedLoadMetricSpecification_statistic :: Lens' CustomizedLoadMetricSpecification MetricStatistic

The statistic of the metric. The only valid value is Sum.

CustomizedScalingMetricSpecification

customizedScalingMetricSpecification_unit :: Lens' CustomizedScalingMetricSpecification (Maybe Text)

The unit of the metric.

customizedScalingMetricSpecification_dimensions :: Lens' CustomizedScalingMetricSpecification (Maybe [MetricDimension])

The dimensions of the metric.

Conditional: If you published your metric with dimensions, you must specify the same dimensions in your customized scaling metric specification.

customizedScalingMetricSpecification_metricName :: Lens' CustomizedScalingMetricSpecification Text

The name of the metric.

customizedScalingMetricSpecification_namespace :: Lens' CustomizedScalingMetricSpecification Text

The namespace of the metric.

customizedScalingMetricSpecification_statistic :: Lens' CustomizedScalingMetricSpecification MetricStatistic

The statistic of the metric.

Datapoint

datapoint_timestamp :: Lens' Datapoint (Maybe UTCTime)

The time stamp for the data point in UTC format.

datapoint_value :: Lens' Datapoint (Maybe Double)

The value of the data point.

MetricDimension

metricDimension_name :: Lens' MetricDimension Text

The name of the dimension.

metricDimension_value :: Lens' MetricDimension Text

The value of the dimension.

PredefinedLoadMetricSpecification

predefinedLoadMetricSpecification_resourceLabel :: Lens' PredefinedLoadMetricSpecification (Maybe Text)

Identifies the resource associated with the metric type. You can't specify a resource label unless the metric type is ALBTargetGroupRequestCount and there is a target group for an Application Load Balancer attached to the Auto Scaling group.

You create the resource label by appending the final portion of the load balancer ARN and the final portion of the target group ARN into a single value, separated by a forward slash (/). The format is app/<load-balancer-name>/<load-balancer-id>/targetgroup/<target-group-name>/<target-group-id>, where:

app/<load-balancer-name>/<load-balancer-id> is the final portion of the load balancer ARN

targetgroup/<target-group-name>/<target-group-id> is the final portion of the target group ARN.

This is an example: app/EC2Co-EcsEl-1TKLTMITMM0EO/f37c06a68c1748aa/targetgroup/EC2Co-Defau-LDNM7Q3ZH1ZN/6d4ea56ca2d6a18d.

To find the ARN for an Application Load Balancer, use the DescribeLoadBalancers API operation. To find the ARN for the target group, use the DescribeTargetGroups API operation.

predefinedLoadMetricSpecification_predefinedLoadMetricType :: Lens' PredefinedLoadMetricSpecification LoadMetricType

The metric type.

PredefinedScalingMetricSpecification

predefinedScalingMetricSpecification_resourceLabel :: Lens' PredefinedScalingMetricSpecification (Maybe Text)

Identifies the resource associated with the metric type. You can't specify a resource label unless the metric type is ALBRequestCountPerTarget and there is a target group for an Application Load Balancer attached to the Auto Scaling group, Spot Fleet request, or ECS service.

You create the resource label by appending the final portion of the load balancer ARN and the final portion of the target group ARN into a single value, separated by a forward slash (/). The format is app/<load-balancer-name>/<load-balancer-id>/targetgroup/<target-group-name>/<target-group-id>, where:

app/<load-balancer-name>/<load-balancer-id> is the final portion of the load balancer ARN

targetgroup/<target-group-name>/<target-group-id> is the final portion of the target group ARN.

This is an example: app/EC2Co-EcsEl-1TKLTMITMM0EO/f37c06a68c1748aa/targetgroup/EC2Co-Defau-LDNM7Q3ZH1ZN/6d4ea56ca2d6a18d.

To find the ARN for an Application Load Balancer, use the DescribeLoadBalancers API operation. To find the ARN for the target group, use the DescribeTargetGroups API operation.

predefinedScalingMetricSpecification_predefinedScalingMetricType :: Lens' PredefinedScalingMetricSpecification ScalingMetricType

The metric type. The ALBRequestCountPerTarget metric type applies only to Auto Scaling groups, Spot Fleet requests, and ECS services.

ScalingInstruction

scalingInstruction_disableDynamicScaling :: Lens' ScalingInstruction (Maybe Bool)

Controls whether dynamic scaling by AWS Auto Scaling is disabled. When dynamic scaling is enabled, AWS Auto Scaling creates target tracking scaling policies based on the specified target tracking configurations.

The default is enabled (false).

scalingInstruction_predefinedLoadMetricSpecification :: Lens' ScalingInstruction (Maybe PredefinedLoadMetricSpecification)

The predefined load metric to use for predictive scaling. This parameter or a CustomizedLoadMetricSpecification is required when configuring predictive scaling, and cannot be used otherwise.

scalingInstruction_predictiveScalingMaxCapacityBehavior :: Lens' ScalingInstruction (Maybe PredictiveScalingMaxCapacityBehavior)

Defines the behavior that should be applied if the forecast capacity approaches or exceeds the maximum capacity specified for the resource. The default value is SetForecastCapacityToMaxCapacity.

The following are possible values:

SetForecastCapacityToMaxCapacity - AWS Auto Scaling cannot scale resource capacity higher than the maximum capacity. The maximum capacity is enforced as a hard limit.

SetMaxCapacityToForecastCapacity - AWS Auto Scaling may scale resource capacity higher than the maximum capacity to equal but not exceed forecast capacity.

SetMaxCapacityAboveForecastCapacity - AWS Auto Scaling may scale resource capacity higher than the maximum capacity by a specified buffer value. The intention is to give the target tracking scaling policy extra capacity if unexpected traffic occurs.

Only valid when configuring predictive scaling.

scalingInstruction_customizedLoadMetricSpecification :: Lens' ScalingInstruction (Maybe CustomizedLoadMetricSpecification)

The customized load metric to use for predictive scaling. This parameter or a PredefinedLoadMetricSpecification is required when configuring predictive scaling, and cannot be used otherwise.

scalingInstruction_predictiveScalingMaxCapacityBuffer :: Lens' ScalingInstruction (Maybe Int)

The size of the capacity buffer to use when the forecast capacity is close to or exceeds the maximum capacity. The value is specified as a percentage relative to the forecast capacity. For example, if the buffer is 10, this means a 10 percent buffer, such that if the forecast capacity is 50, and the maximum capacity is 40, then the effective maximum capacity is 55.

Only valid when configuring predictive scaling. Required if the PredictiveScalingMaxCapacityBehavior is set to SetMaxCapacityAboveForecastCapacity, and cannot be used otherwise.

The range is 1-100.

scalingInstruction_predictiveScalingMode :: Lens' ScalingInstruction (Maybe PredictiveScalingMode)

The predictive scaling mode. The default value is ForecastAndScale. Otherwise, AWS Auto Scaling forecasts capacity but does not create any scheduled scaling actions based on the capacity forecast.

scalingInstruction_scalingPolicyUpdateBehavior :: Lens' ScalingInstruction (Maybe ScalingPolicyUpdateBehavior)

Controls whether a resource's externally created scaling policies are kept or replaced.

The default value is KeepExternalPolicies. If the parameter is set to ReplaceExternalPolicies, any scaling policies that are external to AWS Auto Scaling are deleted and new target tracking scaling policies created.

Only valid when configuring dynamic scaling.

Condition: The number of existing policies to be replaced must be less than or equal to 50. If there are more than 50 policies to be replaced, AWS Auto Scaling keeps all existing policies and does not create new ones.

scalingInstruction_scheduledActionBufferTime :: Lens' ScalingInstruction (Maybe Natural)

The amount of time, in seconds, to buffer the run time of scheduled scaling actions when scaling out. For example, if the forecast says to add capacity at 10:00 AM, and the buffer time is 5 minutes, then the run time of the corresponding scheduled scaling action will be 9:55 AM. The intention is to give resources time to be provisioned. For example, it can take a few minutes to launch an EC2 instance. The actual amount of time required depends on several factors, such as the size of the instance and whether there are startup scripts to complete.

The value must be less than the forecast interval duration of 3600 seconds (60 minutes). The default is 300 seconds.

Only valid when configuring predictive scaling.

scalingInstruction_serviceNamespace :: Lens' ScalingInstruction ServiceNamespace

The namespace of the AWS service.

scalingInstruction_resourceId :: Lens' ScalingInstruction Text

The ID of the resource. This string consists of the resource type and unique identifier.

Auto Scaling group - The resource type is autoScalingGroup and the unique identifier is the name of the Auto Scaling group. Example: autoScalingGroup/my-asg.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

scalingInstruction_scalableDimension :: Lens' ScalingInstruction ScalableDimension

The scalable dimension associated with the resource.

autoscaling:autoScalingGroup:DesiredCapacity - The desired capacity of an Auto Scaling group.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet request.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

scalingInstruction_minCapacity :: Lens' ScalingInstruction Int

The minimum capacity of the resource.

scalingInstruction_maxCapacity :: Lens' ScalingInstruction Int

The maximum capacity of the resource. The exception to this upper limit is if you specify a non-default setting for PredictiveScalingMaxCapacityBehavior.

scalingInstruction_targetTrackingConfigurations :: Lens' ScalingInstruction [TargetTrackingConfiguration]

The target tracking configurations (up to 10). Each of these structures must specify a unique scaling metric and a target value for the metric.

ScalingPlan

scalingPlan_statusMessage :: Lens' ScalingPlan (Maybe Text)

A simple message about the current status of the scaling plan.

scalingPlan_creationTime :: Lens' ScalingPlan (Maybe UTCTime)

The Unix time stamp when the scaling plan was created.

scalingPlan_statusStartTime :: Lens' ScalingPlan (Maybe UTCTime)

The Unix time stamp when the scaling plan entered the current status.

scalingPlan_scalingPlanName :: Lens' ScalingPlan Text

The name of the scaling plan.

scalingPlan_scalingPlanVersion :: Lens' ScalingPlan Integer

The version number of the scaling plan.

scalingPlan_applicationSource :: Lens' ScalingPlan ApplicationSource

A CloudFormation stack or a set of tags. You can create one scaling plan per application source.

scalingPlan_scalingInstructions :: Lens' ScalingPlan [ScalingInstruction]

The scaling instructions.

scalingPlan_statusCode :: Lens' ScalingPlan ScalingPlanStatusCode

The status of the scaling plan.

Active - The scaling plan is active.

ActiveWithProblems - The scaling plan is active, but the scaling configuration for one or more resources could not be applied.

CreationInProgress - The scaling plan is being created.

CreationFailed - The scaling plan could not be created.

DeletionInProgress - The scaling plan is being deleted.

DeletionFailed - The scaling plan could not be deleted.

UpdateInProgress - The scaling plan is being updated.

UpdateFailed - The scaling plan could not be updated.

ScalingPlanResource

scalingPlanResource_scalingPolicies :: Lens' ScalingPlanResource (Maybe [ScalingPolicy])

The scaling policies.

scalingPlanResource_scalingStatusMessage :: Lens' ScalingPlanResource (Maybe Text)

A simple message about the current scaling status of the resource.

scalingPlanResource_scalingPlanName :: Lens' ScalingPlanResource Text

The name of the scaling plan.

scalingPlanResource_scalingPlanVersion :: Lens' ScalingPlanResource Integer

The version number of the scaling plan.

scalingPlanResource_serviceNamespace :: Lens' ScalingPlanResource ServiceNamespace

The namespace of the AWS service.

scalingPlanResource_resourceId :: Lens' ScalingPlanResource Text

The ID of the resource. This string consists of the resource type and unique identifier.

Auto Scaling group - The resource type is autoScalingGroup and the unique identifier is the name of the Auto Scaling group. Example: autoScalingGroup/my-asg.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

scalingPlanResource_scalableDimension :: Lens' ScalingPlanResource ScalableDimension

The scalable dimension for the resource.

autoscaling:autoScalingGroup:DesiredCapacity - The desired capacity of an Auto Scaling group.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet request.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

scalingPlanResource_scalingStatusCode :: Lens' ScalingPlanResource ScalingStatusCode

The scaling status of the resource.

Active - The scaling configuration is active.

Inactive - The scaling configuration is not active because the scaling plan is being created or the scaling configuration could not be applied. Check the status message for more information.

PartiallyActive - The scaling configuration is partially active because the scaling plan is being created or deleted or the scaling configuration could not be fully applied. Check the status message for more information.

ScalingPolicy

scalingPolicy_targetTrackingConfiguration :: Lens' ScalingPolicy (Maybe TargetTrackingConfiguration)

The target tracking scaling policy. Includes support for predefined or customized metrics.

scalingPolicy_policyName :: Lens' ScalingPolicy Text

The name of the scaling policy.

scalingPolicy_policyType :: Lens' ScalingPolicy PolicyType

The type of scaling policy.

TagFilter

tagFilter_key :: Lens' TagFilter (Maybe Text)

The tag key.

tagFilter_values :: Lens' TagFilter (Maybe [Text])

The tag values (0 to 20).

TargetTrackingConfiguration

targetTrackingConfiguration_disableScaleIn :: Lens' TargetTrackingConfiguration (Maybe Bool)

Indicates whether scale in by the target tracking scaling policy is disabled. If the value is true, scale in is disabled and the target tracking scaling policy doesn't remove capacity from the scalable resource. Otherwise, scale in is enabled and the target tracking scaling policy can remove capacity from the scalable resource.

The default value is false.

targetTrackingConfiguration_customizedScalingMetricSpecification :: Lens' TargetTrackingConfiguration (Maybe CustomizedScalingMetricSpecification)

A customized metric. You can specify either a predefined metric or a customized metric.

targetTrackingConfiguration_predefinedScalingMetricSpecification :: Lens' TargetTrackingConfiguration (Maybe PredefinedScalingMetricSpecification)

A predefined metric. You can specify either a predefined metric or a customized metric.

targetTrackingConfiguration_estimatedInstanceWarmup :: Lens' TargetTrackingConfiguration (Maybe Int)

The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics. This value is used only if the resource is an Auto Scaling group.

targetTrackingConfiguration_scaleOutCooldown :: Lens' TargetTrackingConfiguration (Maybe Int)

The amount of time, in seconds, to wait for a previous scale-out activity to take effect. This property is not used if the scalable resource is an Auto Scaling group.

With the scale-out cooldown period, the intention is to continuously (but not excessively) scale out. After Auto Scaling successfully scales out using a target tracking scaling policy, it starts to calculate the cooldown time. The scaling policy won't increase the desired capacity again unless either a larger scale out is triggered or the cooldown period ends.

targetTrackingConfiguration_scaleInCooldown :: Lens' TargetTrackingConfiguration (Maybe Int)

The amount of time, in seconds, after a scale-in activity completes before another scale-in activity can start. This property is not used if the scalable resource is an Auto Scaling group.

With the scale-in cooldown period, the intention is to scale in conservatively to protect your application’s availability, so scale-in activities are blocked until the cooldown period has expired. However, if another alarm triggers a scale-out activity during the scale-in cooldown period, Auto Scaling scales out the target immediately. In this case, the scale-in cooldown period stops and doesn't complete.

targetTrackingConfiguration_targetValue :: Lens' TargetTrackingConfiguration Double

The target value for the metric. Although this property accepts numbers of type Double, it won't accept values that are either too small or too large. Values must be in the range of -2^360 to 2^360.