Network.AWS.AutoScalingPlans.Types.CustomizedScalingMetricSpecification

Description

Documentation

data CustomizedScalingMetricSpecification

Represents a CloudWatch metric of your choosing that can be used for dynamic scaling as part of a target tracking scaling policy.

To create your customized scaling metric specification:

Add values for each required parameter from CloudWatch. You can use an existing metric, or a new metric that you create. To use your own metric, you must first publish the metric to CloudWatch. For more information, see Publish Custom Metrics in the Amazon CloudWatch User Guide.

Choose a metric that changes proportionally with capacity. The value of the metric should increase or decrease in inverse proportion to the number of capacity units. That is, the value of the metric should decrease when capacity increases.

For information about terminology, available metrics, or how to publish new metrics, see Amazon CloudWatch Concepts in the Amazon CloudWatch User Guide.

See: newCustomizedScalingMetricSpecification smart constructor.

Constructors

CustomizedScalingMetricSpecification'

Fields

dimensions :: Maybe [MetricDimension]

The dimensions of the metric.

Conditional: If you published your metric with dimensions, you must specify the same dimensions in your customized scaling metric specification.

unit :: Maybe Text

The unit of the metric.

metricName :: Text

The name of the metric.

namespace :: Text

The namespace of the metric.

statistic :: MetricStatistic

The statistic of the metric.

newCustomizedScalingMetricSpecification

Arguments

:: Text
$sel:metricName:CustomizedScalingMetricSpecification'

-> Text
$sel:namespace:CustomizedScalingMetricSpecification'

-> MetricStatistic
$sel:statistic:CustomizedScalingMetricSpecification'

-> CustomizedScalingMetricSpecification

Create a value of CustomizedScalingMetricSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensions:CustomizedScalingMetricSpecification', customizedScalingMetricSpecification_dimensions - The dimensions of the metric.

Conditional: If you published your metric with dimensions, you must specify the same dimensions in your customized scaling metric specification.

$sel:unit:CustomizedScalingMetricSpecification', customizedScalingMetricSpecification_unit - The unit of the metric.

$sel:metricName:CustomizedScalingMetricSpecification', customizedScalingMetricSpecification_metricName - The name of the metric.

$sel:namespace:CustomizedScalingMetricSpecification', customizedScalingMetricSpecification_namespace - The namespace of the metric.

$sel:statistic:CustomizedScalingMetricSpecification', customizedScalingMetricSpecification_statistic - The statistic of the metric.

customizedScalingMetricSpecification_dimensions :: Lens' CustomizedScalingMetricSpecification (Maybe [MetricDimension])

The dimensions of the metric.

Conditional: If you published your metric with dimensions, you must specify the same dimensions in your customized scaling metric specification.

customizedScalingMetricSpecification_unit :: Lens' CustomizedScalingMetricSpecification (Maybe Text)

The unit of the metric.

customizedScalingMetricSpecification_metricName :: Lens' CustomizedScalingMetricSpecification Text

The name of the metric.

customizedScalingMetricSpecification_namespace :: Lens' CustomizedScalingMetricSpecification Text

The namespace of the metric.

customizedScalingMetricSpecification_statistic :: Lens' CustomizedScalingMetricSpecification MetricStatistic

The statistic of the metric.