Network.AWS.AutoScalingPlans.Types.ScalingPlan

Description

Documentation

data ScalingPlan

Represents a scaling plan.

See: newScalingPlan smart constructor.

Constructors

ScalingPlan'

Fields

statusMessage :: Maybe Text

A simple message about the current status of the scaling plan.

creationTime :: Maybe POSIX

The Unix time stamp when the scaling plan was created.

statusStartTime :: Maybe POSIX

The Unix time stamp when the scaling plan entered the current status.

scalingPlanName :: Text

The name of the scaling plan.

scalingPlanVersion :: Integer

The version number of the scaling plan.

applicationSource :: ApplicationSource

A CloudFormation stack or a set of tags. You can create one scaling plan per application source.

scalingInstructions :: [ScalingInstruction]

The scaling instructions.

statusCode :: ScalingPlanStatusCode

The status of the scaling plan.

Active - The scaling plan is active.

ActiveWithProblems - The scaling plan is active, but the scaling configuration for one or more resources could not be applied.

CreationInProgress - The scaling plan is being created.

CreationFailed - The scaling plan could not be created.

DeletionInProgress - The scaling plan is being deleted.

DeletionFailed - The scaling plan could not be deleted.

UpdateInProgress - The scaling plan is being updated.

UpdateFailed - The scaling plan could not be updated.

newScalingPlan

Arguments

:: Text
$sel:scalingPlanName:ScalingPlan'

-> Integer
$sel:scalingPlanVersion:ScalingPlan'

-> ApplicationSource
$sel:applicationSource:ScalingPlan'

-> ScalingPlanStatusCode
$sel:statusCode:ScalingPlan'

-> ScalingPlan

Create a value of ScalingPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:ScalingPlan', scalingPlan_statusMessage - A simple message about the current status of the scaling plan.

$sel:creationTime:ScalingPlan', scalingPlan_creationTime - The Unix time stamp when the scaling plan was created.

$sel:statusStartTime:ScalingPlan', scalingPlan_statusStartTime - The Unix time stamp when the scaling plan entered the current status.

$sel:scalingPlanName:ScalingPlan', scalingPlan_scalingPlanName - The name of the scaling plan.

$sel:scalingPlanVersion:ScalingPlan', scalingPlan_scalingPlanVersion - The version number of the scaling plan.

$sel:applicationSource:ScalingPlan', scalingPlan_applicationSource - A CloudFormation stack or a set of tags. You can create one scaling plan per application source.

$sel:scalingInstructions:ScalingPlan', scalingPlan_scalingInstructions - The scaling instructions.

$sel:statusCode:ScalingPlan', scalingPlan_statusCode - The status of the scaling plan.

Active - The scaling plan is active.

ActiveWithProblems - The scaling plan is active, but the scaling configuration for one or more resources could not be applied.

CreationInProgress - The scaling plan is being created.

CreationFailed - The scaling plan could not be created.

DeletionInProgress - The scaling plan is being deleted.

DeletionFailed - The scaling plan could not be deleted.

UpdateInProgress - The scaling plan is being updated.

UpdateFailed - The scaling plan could not be updated.

scalingPlan_statusMessage :: Lens' ScalingPlan (Maybe Text)

A simple message about the current status of the scaling plan.

scalingPlan_creationTime :: Lens' ScalingPlan (Maybe UTCTime)

The Unix time stamp when the scaling plan was created.

scalingPlan_statusStartTime :: Lens' ScalingPlan (Maybe UTCTime)

The Unix time stamp when the scaling plan entered the current status.

scalingPlan_scalingPlanName :: Lens' ScalingPlan Text

The name of the scaling plan.

scalingPlan_scalingPlanVersion :: Lens' ScalingPlan Integer

The version number of the scaling plan.

scalingPlan_applicationSource :: Lens' ScalingPlan ApplicationSource

A CloudFormation stack or a set of tags. You can create one scaling plan per application source.

scalingPlan_scalingInstructions :: Lens' ScalingPlan [ScalingInstruction]

The scaling instructions.

scalingPlan_statusCode :: Lens' ScalingPlan ScalingPlanStatusCode

The status of the scaling plan.

Active - The scaling plan is active.

ActiveWithProblems - The scaling plan is active, but the scaling configuration for one or more resources could not be applied.

CreationInProgress - The scaling plan is being created.

CreationFailed - The scaling plan could not be created.

DeletionInProgress - The scaling plan is being deleted.

DeletionFailed - The scaling plan could not be deleted.

UpdateInProgress - The scaling plan is being updated.

UpdateFailed - The scaling plan could not be updated.