Network.AWS.AutoScalingPlans.Types.ScalingPlanResource

Description

Documentation

data ScalingPlanResource

Represents a scalable resource.

See: newScalingPlanResource smart constructor.

Constructors

ScalingPlanResource'

Fields

scalingStatusMessage :: Maybe Text

A simple message about the current scaling status of the resource.

scalingPolicies :: Maybe [ScalingPolicy]

The scaling policies.

scalingPlanName :: Text

The name of the scaling plan.

scalingPlanVersion :: Integer

The version number of the scaling plan.

serviceNamespace :: ServiceNamespace

The namespace of the AWS service.

resourceId :: Text

The ID of the resource. This string consists of the resource type and unique identifier.

Auto Scaling group - The resource type is autoScalingGroup and the unique identifier is the name of the Auto Scaling group. Example: autoScalingGroup/my-asg.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

scalableDimension :: ScalableDimension

The scalable dimension for the resource.

autoscaling:autoScalingGroup:DesiredCapacity - The desired capacity of an Auto Scaling group.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet request.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

scalingStatusCode :: ScalingStatusCode

The scaling status of the resource.

Active - The scaling configuration is active.

Inactive - The scaling configuration is not active because the scaling plan is being created or the scaling configuration could not be applied. Check the status message for more information.

PartiallyActive - The scaling configuration is partially active because the scaling plan is being created or deleted or the scaling configuration could not be fully applied. Check the status message for more information.

newScalingPlanResource

Arguments

:: Text
$sel:scalingPlanName:ScalingPlanResource'

-> Integer
$sel:scalingPlanVersion:ScalingPlanResource'

-> ServiceNamespace
$sel:serviceNamespace:ScalingPlanResource'

-> Text
$sel:resourceId:ScalingPlanResource'

-> ScalableDimension
$sel:scalableDimension:ScalingPlanResource'

-> ScalingStatusCode
$sel:scalingStatusCode:ScalingPlanResource'

-> ScalingPlanResource

Create a value of ScalingPlanResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scalingStatusMessage:ScalingPlanResource', scalingPlanResource_scalingStatusMessage - A simple message about the current scaling status of the resource.

$sel:scalingPolicies:ScalingPlanResource', scalingPlanResource_scalingPolicies - The scaling policies.

$sel:scalingPlanName:ScalingPlanResource', scalingPlanResource_scalingPlanName - The name of the scaling plan.

$sel:scalingPlanVersion:ScalingPlanResource', scalingPlanResource_scalingPlanVersion - The version number of the scaling plan.

$sel:serviceNamespace:ScalingPlanResource', scalingPlanResource_serviceNamespace - The namespace of the AWS service.

$sel:resourceId:ScalingPlanResource', scalingPlanResource_resourceId - The ID of the resource. This string consists of the resource type and unique identifier.

Auto Scaling group - The resource type is autoScalingGroup and the unique identifier is the name of the Auto Scaling group. Example: autoScalingGroup/my-asg.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

$sel:scalableDimension:ScalingPlanResource', scalingPlanResource_scalableDimension - The scalable dimension for the resource.

autoscaling:autoScalingGroup:DesiredCapacity - The desired capacity of an Auto Scaling group.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet request.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

$sel:scalingStatusCode:ScalingPlanResource', scalingPlanResource_scalingStatusCode - The scaling status of the resource.

Active - The scaling configuration is active.

Inactive - The scaling configuration is not active because the scaling plan is being created or the scaling configuration could not be applied. Check the status message for more information.

PartiallyActive - The scaling configuration is partially active because the scaling plan is being created or deleted or the scaling configuration could not be fully applied. Check the status message for more information.

scalingPlanResource_scalingStatusMessage :: Lens' ScalingPlanResource (Maybe Text)

A simple message about the current scaling status of the resource.

scalingPlanResource_scalingPolicies :: Lens' ScalingPlanResource (Maybe [ScalingPolicy])

The scaling policies.

scalingPlanResource_scalingPlanName :: Lens' ScalingPlanResource Text

The name of the scaling plan.

scalingPlanResource_scalingPlanVersion :: Lens' ScalingPlanResource Integer

The version number of the scaling plan.

scalingPlanResource_serviceNamespace :: Lens' ScalingPlanResource ServiceNamespace

The namespace of the AWS service.

scalingPlanResource_resourceId :: Lens' ScalingPlanResource Text

The ID of the resource. This string consists of the resource type and unique identifier.

Auto Scaling group - The resource type is autoScalingGroup and the unique identifier is the name of the Auto Scaling group. Example: autoScalingGroup/my-asg.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot Fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot Fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

Aurora DB cluster - The resource type is cluster and the unique identifier is the cluster name. Example: cluster:my-db-cluster.

scalingPlanResource_scalableDimension :: Lens' ScalingPlanResource ScalableDimension

The scalable dimension for the resource.

autoscaling:autoScalingGroup:DesiredCapacity - The desired capacity of an Auto Scaling group.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot Fleet request.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

rds:cluster:ReadReplicaCount - The count of Aurora Replicas in an Aurora DB cluster. Available for Aurora MySQL-compatible edition and Aurora PostgreSQL-compatible edition.

scalingPlanResource_scalingStatusCode :: Lens' ScalingPlanResource ScalingStatusCode

The scaling status of the resource.

Active - The scaling configuration is active.

Inactive - The scaling configuration is not active because the scaling plan is being created or the scaling configuration could not be applied. Check the status message for more information.

PartiallyActive - The scaling configuration is partially active because the scaling plan is being created or deleted or the scaling configuration could not be fully applied. Check the status message for more information.