Network.AWS.Backup.CreateBackupPlan

Description

Creates a backup plan using a backup plan name and backup rules. A backup plan is a document that contains information that Backup uses to schedule tasks that create recovery points for resources.

If you call CreateBackupPlan with a plan that already exists, you receive an AlreadyExistsException exception.

Creating a Request

data CreateBackupPlan

See: newCreateBackupPlan smart constructor.

Constructors

CreateBackupPlan'

Fields

backupPlanTags :: Maybe (Sensitive (HashMap Text Text))

To help organize your resources, you can assign your own metadata to the resources that you create. Each tag is a key-value pair. The specified tags are assigned to all backups created with this plan.

creatorRequestId :: Maybe Text

Identifies the request and allows failed requests to be retried without the risk of running the operation twice. If the request includes a CreatorRequestId that matches an existing backup plan, that plan is returned. This parameter is optional.

backupPlan :: BackupPlanInput

Specifies the body of a backup plan. Includes a BackupPlanName and one or more sets of Rules.

newCreateBackupPlan

Arguments

:: BackupPlanInput
$sel:backupPlan:CreateBackupPlan'

-> CreateBackupPlan

Create a value of CreateBackupPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupPlanTags:CreateBackupPlan', createBackupPlan_backupPlanTags - To help organize your resources, you can assign your own metadata to the resources that you create. Each tag is a key-value pair. The specified tags are assigned to all backups created with this plan.

$sel:creatorRequestId:CreateBackupPlan', createBackupPlan_creatorRequestId - Identifies the request and allows failed requests to be retried without the risk of running the operation twice. If the request includes a CreatorRequestId that matches an existing backup plan, that plan is returned. This parameter is optional.

$sel:backupPlan:CreateBackupPlan', createBackupPlan_backupPlan - Specifies the body of a backup plan. Includes a BackupPlanName and one or more sets of Rules.

Request Lenses

createBackupPlan_backupPlanTags :: Lens' CreateBackupPlan (Maybe (HashMap Text Text))

To help organize your resources, you can assign your own metadata to the resources that you create. Each tag is a key-value pair. The specified tags are assigned to all backups created with this plan.

createBackupPlan_creatorRequestId :: Lens' CreateBackupPlan (Maybe Text)

Identifies the request and allows failed requests to be retried without the risk of running the operation twice. If the request includes a CreatorRequestId that matches an existing backup plan, that plan is returned. This parameter is optional.

createBackupPlan_backupPlan :: Lens' CreateBackupPlan BackupPlanInput

Specifies the body of a backup plan. Includes a BackupPlanName and one or more sets of Rules.

Destructuring the Response

data CreateBackupPlanResponse

See: newCreateBackupPlanResponse smart constructor.

Constructors

CreateBackupPlanResponse'

Fields

versionId :: Maybe Text

Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. They cannot be edited.

advancedBackupSettings :: Maybe [AdvancedBackupSetting]

A list of BackupOptions settings for a resource type. This option is only available for Windows Volume Shadow Copy Service (VSS) backup jobs.

backupPlanId :: Maybe Text

Uniquely identifies a backup plan.

backupPlanArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

creationDate :: Maybe POSIX

The date and time that a backup plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

httpStatus :: Int

The response's http status code.

newCreateBackupPlanResponse

Arguments

:: Int
$sel:httpStatus:CreateBackupPlanResponse'

-> CreateBackupPlanResponse

Create a value of CreateBackupPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:CreateBackupPlanResponse', createBackupPlanResponse_versionId - Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. They cannot be edited.

$sel:advancedBackupSettings:CreateBackupPlanResponse', createBackupPlanResponse_advancedBackupSettings - A list of BackupOptions settings for a resource type. This option is only available for Windows Volume Shadow Copy Service (VSS) backup jobs.

$sel:backupPlanId:CreateBackupPlanResponse', createBackupPlanResponse_backupPlanId - Uniquely identifies a backup plan.

$sel:backupPlanArn:CreateBackupPlanResponse', createBackupPlanResponse_backupPlanArn - An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

$sel:creationDate:CreateBackupPlanResponse', createBackupPlanResponse_creationDate - The date and time that a backup plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:httpStatus:CreateBackupPlanResponse', createBackupPlanResponse_httpStatus - The response's http status code.

Response Lenses

createBackupPlanResponse_versionId :: Lens' CreateBackupPlanResponse (Maybe Text)

Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. They cannot be edited.

createBackupPlanResponse_advancedBackupSettings :: Lens' CreateBackupPlanResponse (Maybe [AdvancedBackupSetting])

A list of BackupOptions settings for a resource type. This option is only available for Windows Volume Shadow Copy Service (VSS) backup jobs.

createBackupPlanResponse_backupPlanId :: Lens' CreateBackupPlanResponse (Maybe Text)

Uniquely identifies a backup plan.

createBackupPlanResponse_backupPlanArn :: Lens' CreateBackupPlanResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

createBackupPlanResponse_creationDate :: Lens' CreateBackupPlanResponse (Maybe UTCTime)

The date and time that a backup plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

createBackupPlanResponse_httpStatus :: Lens' CreateBackupPlanResponse Int

The response's http status code.