Network.AWS.Backup.CreateFramework

Description

Creates a framework with one or more controls. A framework is a collection of controls that you can use to evaluate your backup practices. By using pre-built customizable controls to define your policies, you can evaluate whether your backup practices comply with your policies and which resources are not yet in compliance.

Creating a Request

data CreateFramework

See: newCreateFramework smart constructor.

Constructors

CreateFramework'

Fields

idempotencyToken :: Maybe Text

A customer-chosen string that you can use to distinguish between otherwise identical calls to CreateFrameworkInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

frameworkDescription :: Maybe Text

An optional description of the framework with a maximum of 1,024 characters.

frameworkTags :: Maybe (HashMap Text Text)

Metadata that you can assign to help organize the frameworks that you create. Each tag is a key-value pair.

frameworkName :: Text

The unique name of the framework. The name must be between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

frameworkControls :: [FrameworkControl]

A list of the controls that make up the framework. Each control in the list has a name, input parameters, and scope.

newCreateFramework

Arguments

:: Text
$sel:frameworkName:CreateFramework'

-> CreateFramework

Create a value of CreateFramework with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idempotencyToken:CreateFramework', createFramework_idempotencyToken - A customer-chosen string that you can use to distinguish between otherwise identical calls to CreateFrameworkInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

$sel:frameworkDescription:CreateFramework', createFramework_frameworkDescription - An optional description of the framework with a maximum of 1,024 characters.

$sel:frameworkTags:CreateFramework', createFramework_frameworkTags - Metadata that you can assign to help organize the frameworks that you create. Each tag is a key-value pair.

$sel:frameworkName:CreateFramework', createFramework_frameworkName - The unique name of the framework. The name must be between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

$sel:frameworkControls:CreateFramework', createFramework_frameworkControls - A list of the controls that make up the framework. Each control in the list has a name, input parameters, and scope.

Request Lenses

createFramework_idempotencyToken :: Lens' CreateFramework (Maybe Text)

A customer-chosen string that you can use to distinguish between otherwise identical calls to CreateFrameworkInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

createFramework_frameworkDescription :: Lens' CreateFramework (Maybe Text)

An optional description of the framework with a maximum of 1,024 characters.

createFramework_frameworkTags :: Lens' CreateFramework (Maybe (HashMap Text Text))

Metadata that you can assign to help organize the frameworks that you create. Each tag is a key-value pair.

createFramework_frameworkName :: Lens' CreateFramework Text

The unique name of the framework. The name must be between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

createFramework_frameworkControls :: Lens' CreateFramework [FrameworkControl]

A list of the controls that make up the framework. Each control in the list has a name, input parameters, and scope.

Destructuring the Response

data CreateFrameworkResponse

See: newCreateFrameworkResponse smart constructor.

Constructors

CreateFrameworkResponse'

Fields

frameworkArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

frameworkName :: Maybe Text

The unique name of the framework. The name must be between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

httpStatus :: Int

The response's http status code.

newCreateFrameworkResponse

Arguments

:: Int
$sel:httpStatus:CreateFrameworkResponse'

-> CreateFrameworkResponse

Create a value of CreateFrameworkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:frameworkArn:CreateFrameworkResponse', createFrameworkResponse_frameworkArn - An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

$sel:frameworkName:CreateFramework', createFrameworkResponse_frameworkName - The unique name of the framework. The name must be between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

$sel:httpStatus:CreateFrameworkResponse', createFrameworkResponse_httpStatus - The response's http status code.

Response Lenses

createFrameworkResponse_frameworkArn :: Lens' CreateFrameworkResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

createFrameworkResponse_frameworkName :: Lens' CreateFrameworkResponse (Maybe Text)

The unique name of the framework. The name must be between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

createFrameworkResponse_httpStatus :: Lens' CreateFrameworkResponse Int

The response's http status code.