Network.AWS.Backup.CreateReportPlan

Description

Creates a report plan. A report plan is a document that contains information about the contents of the report and where Backup will deliver it.

If you call CreateReportPlan with a plan that already exists, you receive an AlreadyExistsException exception.

Creating a Request

data CreateReportPlan

See: newCreateReportPlan smart constructor.

Constructors

CreateReportPlan'

Fields

idempotencyToken :: Maybe Text

A customer-chosen string that you can use to distinguish between otherwise identical calls to CreateReportPlanInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

reportPlanTags :: Maybe (HashMap Text Text)

Metadata that you can assign to help organize the report plans that you create. Each tag is a key-value pair.

reportPlanDescription :: Maybe Text

An optional description of the report plan with a maximum of 1,024 characters.

reportPlanName :: Text

The unique name of the report plan. The name must be between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

reportDeliveryChannel :: ReportDeliveryChannel

A structure that contains information about where and how to deliver your reports, specifically your Amazon S3 bucket name, S3 key prefix, and the formats of your reports.

reportSetting :: ReportSetting

Identifies the report template for the report. Reports are built using a report template. The report templates are:

RESOURCE_COMPLIANCE_REPORT | CONTROL_COMPLIANCE_REPORT | BACKUP_JOB_REPORT | COPY_JOB_REPORT | RESTORE_JOB_REPORT

If the report template is RESOURCE_COMPLIANCE_REPORT or CONTROL_COMPLIANCE_REPORT, this API resource also describes the report coverage by Amazon Web Services Regions and frameworks.

newCreateReportPlan

Arguments

:: Text
$sel:reportPlanName:CreateReportPlan'

-> ReportDeliveryChannel
$sel:reportDeliveryChannel:CreateReportPlan'

-> ReportSetting
$sel:reportSetting:CreateReportPlan'

-> CreateReportPlan

Create a value of CreateReportPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idempotencyToken:CreateReportPlan', createReportPlan_idempotencyToken - A customer-chosen string that you can use to distinguish between otherwise identical calls to CreateReportPlanInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

$sel:reportPlanTags:CreateReportPlan', createReportPlan_reportPlanTags - Metadata that you can assign to help organize the report plans that you create. Each tag is a key-value pair.

$sel:reportPlanDescription:CreateReportPlan', createReportPlan_reportPlanDescription - An optional description of the report plan with a maximum of 1,024 characters.

$sel:reportPlanName:CreateReportPlan', createReportPlan_reportPlanName - The unique name of the report plan. The name must be between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

$sel:reportDeliveryChannel:CreateReportPlan', createReportPlan_reportDeliveryChannel - A structure that contains information about where and how to deliver your reports, specifically your Amazon S3 bucket name, S3 key prefix, and the formats of your reports.

$sel:reportSetting:CreateReportPlan', createReportPlan_reportSetting - Identifies the report template for the report. Reports are built using a report template. The report templates are:

RESOURCE_COMPLIANCE_REPORT | CONTROL_COMPLIANCE_REPORT | BACKUP_JOB_REPORT | COPY_JOB_REPORT | RESTORE_JOB_REPORT

If the report template is RESOURCE_COMPLIANCE_REPORT or CONTROL_COMPLIANCE_REPORT, this API resource also describes the report coverage by Amazon Web Services Regions and frameworks.

Request Lenses

createReportPlan_idempotencyToken :: Lens' CreateReportPlan (Maybe Text)

A customer-chosen string that you can use to distinguish between otherwise identical calls to CreateReportPlanInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

createReportPlan_reportPlanTags :: Lens' CreateReportPlan (Maybe (HashMap Text Text))

Metadata that you can assign to help organize the report plans that you create. Each tag is a key-value pair.

createReportPlan_reportPlanDescription :: Lens' CreateReportPlan (Maybe Text)

An optional description of the report plan with a maximum of 1,024 characters.

createReportPlan_reportPlanName :: Lens' CreateReportPlan Text

The unique name of the report plan. The name must be between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

createReportPlan_reportDeliveryChannel :: Lens' CreateReportPlan ReportDeliveryChannel

A structure that contains information about where and how to deliver your reports, specifically your Amazon S3 bucket name, S3 key prefix, and the formats of your reports.

createReportPlan_reportSetting :: Lens' CreateReportPlan ReportSetting

Identifies the report template for the report. Reports are built using a report template. The report templates are:

RESOURCE_COMPLIANCE_REPORT | CONTROL_COMPLIANCE_REPORT | BACKUP_JOB_REPORT | COPY_JOB_REPORT | RESTORE_JOB_REPORT

If the report template is RESOURCE_COMPLIANCE_REPORT or CONTROL_COMPLIANCE_REPORT, this API resource also describes the report coverage by Amazon Web Services Regions and frameworks.

Destructuring the Response

data CreateReportPlanResponse

See: newCreateReportPlanResponse smart constructor.

Constructors

CreateReportPlanResponse'

Fields

creationTime :: Maybe POSIX

The date and time a backup vault is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

reportPlanName :: Maybe Text

The unique name of the report plan.

reportPlanArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

httpStatus :: Int

The response's http status code.

newCreateReportPlanResponse

Arguments

:: Int
$sel:httpStatus:CreateReportPlanResponse'

-> CreateReportPlanResponse

Create a value of CreateReportPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CreateReportPlanResponse', createReportPlanResponse_creationTime - The date and time a backup vault is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:reportPlanName:CreateReportPlan', createReportPlanResponse_reportPlanName - The unique name of the report plan.

$sel:reportPlanArn:CreateReportPlanResponse', createReportPlanResponse_reportPlanArn - An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

$sel:httpStatus:CreateReportPlanResponse', createReportPlanResponse_httpStatus - The response's http status code.

Response Lenses

createReportPlanResponse_creationTime :: Lens' CreateReportPlanResponse (Maybe UTCTime)

The date and time a backup vault is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

createReportPlanResponse_reportPlanName :: Lens' CreateReportPlanResponse (Maybe Text)

The unique name of the report plan.

createReportPlanResponse_reportPlanArn :: Lens' CreateReportPlanResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

createReportPlanResponse_httpStatus :: Lens' CreateReportPlanResponse Int

The response's http status code.