Network.AWS.Backup.DeleteBackupPlan

Description

Deletes a backup plan. A backup plan can only be deleted after all associated selections of resources have been deleted. Deleting a backup plan deletes the current version of a backup plan. Previous versions, if any, will still exist.

Creating a Request

data DeleteBackupPlan

See: newDeleteBackupPlan smart constructor.

Constructors

DeleteBackupPlan'

Fields

backupPlanId :: Text

Uniquely identifies a backup plan.

newDeleteBackupPlan

Arguments

:: Text
$sel:backupPlanId:DeleteBackupPlan'

-> DeleteBackupPlan

Create a value of DeleteBackupPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupPlanId:DeleteBackupPlan', deleteBackupPlan_backupPlanId - Uniquely identifies a backup plan.

Request Lenses

deleteBackupPlan_backupPlanId :: Lens' DeleteBackupPlan Text

Uniquely identifies a backup plan.

Destructuring the Response

data DeleteBackupPlanResponse

See: newDeleteBackupPlanResponse smart constructor.

Constructors

DeleteBackupPlanResponse'

Fields

versionId :: Maybe Text

Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

backupPlanId :: Maybe Text

Uniquely identifies a backup plan.

backupPlanArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

deletionDate :: Maybe POSIX

The date and time a backup plan is deleted, in Unix format and Coordinated Universal Time (UTC). The value of DeletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

httpStatus :: Int

The response's http status code.

newDeleteBackupPlanResponse

Arguments

:: Int
$sel:httpStatus:DeleteBackupPlanResponse'

-> DeleteBackupPlanResponse

Create a value of DeleteBackupPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:DeleteBackupPlanResponse', deleteBackupPlanResponse_versionId - Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

$sel:backupPlanId:DeleteBackupPlan', deleteBackupPlanResponse_backupPlanId - Uniquely identifies a backup plan.

$sel:backupPlanArn:DeleteBackupPlanResponse', deleteBackupPlanResponse_backupPlanArn - An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

$sel:deletionDate:DeleteBackupPlanResponse', deleteBackupPlanResponse_deletionDate - The date and time a backup plan is deleted, in Unix format and Coordinated Universal Time (UTC). The value of DeletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:httpStatus:DeleteBackupPlanResponse', deleteBackupPlanResponse_httpStatus - The response's http status code.

Response Lenses

deleteBackupPlanResponse_versionId :: Lens' DeleteBackupPlanResponse (Maybe Text)

Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

deleteBackupPlanResponse_backupPlanId :: Lens' DeleteBackupPlanResponse (Maybe Text)

Uniquely identifies a backup plan.

deleteBackupPlanResponse_backupPlanArn :: Lens' DeleteBackupPlanResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

deleteBackupPlanResponse_deletionDate :: Lens' DeleteBackupPlanResponse (Maybe UTCTime)

The date and time a backup plan is deleted, in Unix format and Coordinated Universal Time (UTC). The value of DeletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

deleteBackupPlanResponse_httpStatus :: Lens' DeleteBackupPlanResponse Int

The response's http status code.