Network.AWS.Backup.DescribeFramework

Description

Returns the framework details for the specified FrameworkName.

Creating a Request

data DescribeFramework

See: newDescribeFramework smart constructor.

Constructors

DescribeFramework'

Fields

frameworkName :: Text

The unique name of a framework.

newDescribeFramework

Arguments

:: Text
$sel:frameworkName:DescribeFramework'

-> DescribeFramework

Create a value of DescribeFramework with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:frameworkName:DescribeFramework', describeFramework_frameworkName - The unique name of a framework.

Request Lenses

describeFramework_frameworkName :: Lens' DescribeFramework Text

The unique name of a framework.

Destructuring the Response

data DescribeFrameworkResponse

See: newDescribeFrameworkResponse smart constructor.

Constructors

DescribeFrameworkResponse'

Fields

creationTime :: Maybe POSIX

The date and time that a framework is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

idempotencyToken :: Maybe Text

A customer-chosen string that you can use to distinguish between otherwise identical calls to DescribeFrameworkOutput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

frameworkDescription :: Maybe Text

An optional description of the framework.

frameworkStatus :: Maybe Text

A framework consists of one or more controls. Each control governs a resource, such as backup plans, backup selections, backup vaults, or recovery points. You can also turn Config recording on or off for each resource. The statuses are:

ACTIVE when recording is turned on for all resources governed by the framework.

PARTIALLY_ACTIVE when recording is turned off for at least one resource governed by the framework.

INACTIVE when recording is turned off for all resources governed by the framework.

UNAVAILABLE when Backup is unable to validate recording status at this time.

frameworkControls :: Maybe [FrameworkControl]

A list of the controls that make up the framework. Each control in the list has a name, input parameters, and scope.

frameworkArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

deploymentStatus :: Maybe Text

The deployment status of a framework. The statuses are:

CREATE_IN_PROGRESS | UPDATE_IN_PROGRESS | DELETE_IN_PROGRESS | COMPLETED | FAILED

frameworkName :: Maybe Text

The unique name of a framework.

httpStatus :: Int

The response's http status code.

newDescribeFrameworkResponse

Arguments

:: Int
$sel:httpStatus:DescribeFrameworkResponse'

-> DescribeFrameworkResponse

Create a value of DescribeFrameworkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeFrameworkResponse', describeFrameworkResponse_creationTime - The date and time that a framework is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:idempotencyToken:DescribeFrameworkResponse', describeFrameworkResponse_idempotencyToken - A customer-chosen string that you can use to distinguish between otherwise identical calls to DescribeFrameworkOutput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

$sel:frameworkDescription:DescribeFrameworkResponse', describeFrameworkResponse_frameworkDescription - An optional description of the framework.

$sel:frameworkStatus:DescribeFrameworkResponse', describeFrameworkResponse_frameworkStatus - A framework consists of one or more controls. Each control governs a resource, such as backup plans, backup selections, backup vaults, or recovery points. You can also turn Config recording on or off for each resource. The statuses are:

ACTIVE when recording is turned on for all resources governed by the framework.

PARTIALLY_ACTIVE when recording is turned off for at least one resource governed by the framework.

INACTIVE when recording is turned off for all resources governed by the framework.

UNAVAILABLE when Backup is unable to validate recording status at this time.

$sel:frameworkControls:DescribeFrameworkResponse', describeFrameworkResponse_frameworkControls - A list of the controls that make up the framework. Each control in the list has a name, input parameters, and scope.

$sel:frameworkArn:DescribeFrameworkResponse', describeFrameworkResponse_frameworkArn - An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

$sel:deploymentStatus:DescribeFrameworkResponse', describeFrameworkResponse_deploymentStatus - The deployment status of a framework. The statuses are:

CREATE_IN_PROGRESS | UPDATE_IN_PROGRESS | DELETE_IN_PROGRESS | COMPLETED | FAILED

$sel:frameworkName:DescribeFramework', describeFrameworkResponse_frameworkName - The unique name of a framework.

$sel:httpStatus:DescribeFrameworkResponse', describeFrameworkResponse_httpStatus - The response's http status code.

Response Lenses

describeFrameworkResponse_creationTime :: Lens' DescribeFrameworkResponse (Maybe UTCTime)

The date and time that a framework is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

describeFrameworkResponse_idempotencyToken :: Lens' DescribeFrameworkResponse (Maybe Text)

A customer-chosen string that you can use to distinguish between otherwise identical calls to DescribeFrameworkOutput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

describeFrameworkResponse_frameworkDescription :: Lens' DescribeFrameworkResponse (Maybe Text)

An optional description of the framework.

describeFrameworkResponse_frameworkStatus :: Lens' DescribeFrameworkResponse (Maybe Text)

A framework consists of one or more controls. Each control governs a resource, such as backup plans, backup selections, backup vaults, or recovery points. You can also turn Config recording on or off for each resource. The statuses are:

ACTIVE when recording is turned on for all resources governed by the framework.

PARTIALLY_ACTIVE when recording is turned off for at least one resource governed by the framework.

INACTIVE when recording is turned off for all resources governed by the framework.

UNAVAILABLE when Backup is unable to validate recording status at this time.

describeFrameworkResponse_frameworkControls :: Lens' DescribeFrameworkResponse (Maybe [FrameworkControl])

A list of the controls that make up the framework. Each control in the list has a name, input parameters, and scope.

describeFrameworkResponse_frameworkArn :: Lens' DescribeFrameworkResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

describeFrameworkResponse_deploymentStatus :: Lens' DescribeFrameworkResponse (Maybe Text)

The deployment status of a framework. The statuses are:

CREATE_IN_PROGRESS | UPDATE_IN_PROGRESS | DELETE_IN_PROGRESS | COMPLETED | FAILED

describeFrameworkResponse_frameworkName :: Lens' DescribeFrameworkResponse (Maybe Text)

The unique name of a framework.

describeFrameworkResponse_httpStatus :: Lens' DescribeFrameworkResponse Int

The response's http status code.