Network.AWS.Backup.DescribeGlobalSettings

Description

Describes whether the Amazon Web Services account is opted in to cross-account backup. Returns an error if the account is not a member of an Organizations organization. Example: describe-global-settings --region us-west-2

Creating a Request

data DescribeGlobalSettings

See: newDescribeGlobalSettings smart constructor.

Constructors

DescribeGlobalSettings'

newDescribeGlobalSettings :: DescribeGlobalSettings

Create a value of DescribeGlobalSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeGlobalSettingsResponse

See: newDescribeGlobalSettingsResponse smart constructor.

Constructors

DescribeGlobalSettingsResponse'

Fields

globalSettings :: Maybe (HashMap Text Text)

The status of the flag isCrossAccountBackupEnabled.

lastUpdateTime :: Maybe POSIX

The date and time that the flag isCrossAccountBackupEnabled was last updated. This update is in Unix format and Coordinated Universal Time (UTC). The value of LastUpdateTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

httpStatus :: Int

The response's http status code.

newDescribeGlobalSettingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeGlobalSettingsResponse'

-> DescribeGlobalSettingsResponse

Create a value of DescribeGlobalSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalSettings:DescribeGlobalSettingsResponse', describeGlobalSettingsResponse_globalSettings - The status of the flag isCrossAccountBackupEnabled.

$sel:lastUpdateTime:DescribeGlobalSettingsResponse', describeGlobalSettingsResponse_lastUpdateTime - The date and time that the flag isCrossAccountBackupEnabled was last updated. This update is in Unix format and Coordinated Universal Time (UTC). The value of LastUpdateTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:httpStatus:DescribeGlobalSettingsResponse', describeGlobalSettingsResponse_httpStatus - The response's http status code.

Response Lenses

describeGlobalSettingsResponse_globalSettings :: Lens' DescribeGlobalSettingsResponse (Maybe (HashMap Text Text))

The status of the flag isCrossAccountBackupEnabled.

describeGlobalSettingsResponse_lastUpdateTime :: Lens' DescribeGlobalSettingsResponse (Maybe UTCTime)

The date and time that the flag isCrossAccountBackupEnabled was last updated. This update is in Unix format and Coordinated Universal Time (UTC). The value of LastUpdateTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

describeGlobalSettingsResponse_httpStatus :: Lens' DescribeGlobalSettingsResponse Int

The response's http status code.