Network.AWS.Backup.DescribeRecoveryPoint

Description

Returns metadata associated with a recovery point, including ID, status, encryption, and lifecycle.

Creating a Request

data DescribeRecoveryPoint

See: newDescribeRecoveryPoint smart constructor.

Constructors

DescribeRecoveryPoint'

Fields

backupVaultName :: Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

recoveryPointArn :: Text

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

newDescribeRecoveryPoint

Arguments

:: Text
$sel:backupVaultName:DescribeRecoveryPoint'

-> Text
$sel:recoveryPointArn:DescribeRecoveryPoint'

-> DescribeRecoveryPoint

Create a value of DescribeRecoveryPoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupVaultName:DescribeRecoveryPoint', describeRecoveryPoint_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

$sel:recoveryPointArn:DescribeRecoveryPoint', describeRecoveryPoint_recoveryPointArn - An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

Request Lenses

describeRecoveryPoint_backupVaultName :: Lens' DescribeRecoveryPoint Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

describeRecoveryPoint_recoveryPointArn :: Lens' DescribeRecoveryPoint Text

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

Destructuring the Response

data DescribeRecoveryPointResponse

See: newDescribeRecoveryPointResponse smart constructor.

Constructors

DescribeRecoveryPointResponse'

Fields

isEncrypted :: Maybe Bool

A Boolean value that is returned as TRUE if the specified recovery point is encrypted, or FALSE if the recovery point is not encrypted.

status :: Maybe RecoveryPointStatus

A status code specifying the state of the recovery point.

PARTIAL status indicates Backup could not create the recovery point before the backup window closed. To increase your backup plan window using the API, see UpdateBackupPlan. You can also increase your backup plan window using the Console by choosing and editing your backup plan.

EXPIRED status indicates that the recovery point has exceeded its retention period, but Backup lacks permission or is otherwise unable to delete it. To manually delete these recovery points, see Step 3: Delete the recovery points in the Clean up resources section of Getting started.

iamRoleArn :: Maybe Text

Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

resourceType :: Maybe Text

The type of Amazon Web Services resource to save as a recovery point; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database.

createdBy :: Maybe RecoveryPointCreator

Contains identifying information about the creation of a recovery point, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan used to create it.

calculatedLifecycle :: Maybe CalculatedLifecycle

A CalculatedLifecycle object containing DeleteAt and MoveToColdStorageAt timestamps.

lifecycle :: Maybe Lifecycle

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups that are transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

backupVaultArn :: Maybe Text

An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

sourceBackupVaultArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies the source vault where the resource was originally backed up in; for example, arn:aws:backup:us-east-1:123456789012:vault:BackupVault. If the recovery is restored to the same Amazon Web Services account or Region, this value will be null.

lastRestoreTime :: Maybe POSIX

The date and time that a recovery point was last restored, in Unix format and Coordinated Universal Time (UTC). The value of LastRestoreTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

resourceArn :: Maybe Text

An ARN that uniquely identifies a saved resource. The format of the ARN depends on the resource type.

statusMessage :: Maybe Text

A status message explaining the reason for the recovery point deletion failure.

storageClass :: Maybe StorageClass

Specifies the storage class of the recovery point. Valid values are WARM or COLD.

recoveryPointArn :: Maybe Text

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

encryptionKeyArn :: Maybe Text

The server-side encryption key used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

backupSizeInBytes :: Maybe Integer

The size, in bytes, of a backup.

creationDate :: Maybe POSIX

The date and time that a recovery point is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

completionDate :: Maybe POSIX

The date and time that a job to create a recovery point is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupVaultName :: Maybe Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Region where they are created. They consist of lowercase letters, numbers, and hyphens.

httpStatus :: Int

The response's http status code.

newDescribeRecoveryPointResponse

Arguments

:: Int
$sel:httpStatus:DescribeRecoveryPointResponse'

-> DescribeRecoveryPointResponse

Create a value of DescribeRecoveryPointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isEncrypted:DescribeRecoveryPointResponse', describeRecoveryPointResponse_isEncrypted - A Boolean value that is returned as TRUE if the specified recovery point is encrypted, or FALSE if the recovery point is not encrypted.

$sel:status:DescribeRecoveryPointResponse', describeRecoveryPointResponse_status - A status code specifying the state of the recovery point.

PARTIAL status indicates Backup could not create the recovery point before the backup window closed. To increase your backup plan window using the API, see UpdateBackupPlan. You can also increase your backup plan window using the Console by choosing and editing your backup plan.

EXPIRED status indicates that the recovery point has exceeded its retention period, but Backup lacks permission or is otherwise unable to delete it. To manually delete these recovery points, see Step 3: Delete the recovery points in the Clean up resources section of Getting started.

$sel:iamRoleArn:DescribeRecoveryPointResponse', describeRecoveryPointResponse_iamRoleArn - Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

$sel:resourceType:DescribeRecoveryPointResponse', describeRecoveryPointResponse_resourceType - The type of Amazon Web Services resource to save as a recovery point; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database.

$sel:createdBy:DescribeRecoveryPointResponse', describeRecoveryPointResponse_createdBy - Contains identifying information about the creation of a recovery point, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan used to create it.

$sel:calculatedLifecycle:DescribeRecoveryPointResponse', describeRecoveryPointResponse_calculatedLifecycle - A CalculatedLifecycle object containing DeleteAt and MoveToColdStorageAt timestamps.

$sel:lifecycle:DescribeRecoveryPointResponse', describeRecoveryPointResponse_lifecycle - The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups that are transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

$sel:backupVaultArn:DescribeRecoveryPointResponse', describeRecoveryPointResponse_backupVaultArn - An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:sourceBackupVaultArn:DescribeRecoveryPointResponse', describeRecoveryPointResponse_sourceBackupVaultArn - An Amazon Resource Name (ARN) that uniquely identifies the source vault where the resource was originally backed up in; for example, arn:aws:backup:us-east-1:123456789012:vault:BackupVault. If the recovery is restored to the same Amazon Web Services account or Region, this value will be null.

$sel:lastRestoreTime:DescribeRecoveryPointResponse', describeRecoveryPointResponse_lastRestoreTime - The date and time that a recovery point was last restored, in Unix format and Coordinated Universal Time (UTC). The value of LastRestoreTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:resourceArn:DescribeRecoveryPointResponse', describeRecoveryPointResponse_resourceArn - An ARN that uniquely identifies a saved resource. The format of the ARN depends on the resource type.

$sel:statusMessage:DescribeRecoveryPointResponse', describeRecoveryPointResponse_statusMessage - A status message explaining the reason for the recovery point deletion failure.

$sel:storageClass:DescribeRecoveryPointResponse', describeRecoveryPointResponse_storageClass - Specifies the storage class of the recovery point. Valid values are WARM or COLD.

$sel:recoveryPointArn:DescribeRecoveryPoint', describeRecoveryPointResponse_recoveryPointArn - An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:encryptionKeyArn:DescribeRecoveryPointResponse', describeRecoveryPointResponse_encryptionKeyArn - The server-side encryption key used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

$sel:backupSizeInBytes:DescribeRecoveryPointResponse', describeRecoveryPointResponse_backupSizeInBytes - The size, in bytes, of a backup.

$sel:creationDate:DescribeRecoveryPointResponse', describeRecoveryPointResponse_creationDate - The date and time that a recovery point is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:completionDate:DescribeRecoveryPointResponse', describeRecoveryPointResponse_completionDate - The date and time that a job to create a recovery point is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:backupVaultName:DescribeRecoveryPoint', describeRecoveryPointResponse_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Region where they are created. They consist of lowercase letters, numbers, and hyphens.

$sel:httpStatus:DescribeRecoveryPointResponse', describeRecoveryPointResponse_httpStatus - The response's http status code.

Response Lenses

describeRecoveryPointResponse_isEncrypted :: Lens' DescribeRecoveryPointResponse (Maybe Bool)

A Boolean value that is returned as TRUE if the specified recovery point is encrypted, or FALSE if the recovery point is not encrypted.

describeRecoveryPointResponse_status :: Lens' DescribeRecoveryPointResponse (Maybe RecoveryPointStatus)

A status code specifying the state of the recovery point.

PARTIAL status indicates Backup could not create the recovery point before the backup window closed. To increase your backup plan window using the API, see UpdateBackupPlan. You can also increase your backup plan window using the Console by choosing and editing your backup plan.

EXPIRED status indicates that the recovery point has exceeded its retention period, but Backup lacks permission or is otherwise unable to delete it. To manually delete these recovery points, see Step 3: Delete the recovery points in the Clean up resources section of Getting started.

describeRecoveryPointResponse_iamRoleArn :: Lens' DescribeRecoveryPointResponse (Maybe Text)

Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

describeRecoveryPointResponse_resourceType :: Lens' DescribeRecoveryPointResponse (Maybe Text)

The type of Amazon Web Services resource to save as a recovery point; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database.

describeRecoveryPointResponse_createdBy :: Lens' DescribeRecoveryPointResponse (Maybe RecoveryPointCreator)

Contains identifying information about the creation of a recovery point, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan used to create it.

describeRecoveryPointResponse_calculatedLifecycle :: Lens' DescribeRecoveryPointResponse (Maybe CalculatedLifecycle)

A CalculatedLifecycle object containing DeleteAt and MoveToColdStorageAt timestamps.

describeRecoveryPointResponse_lifecycle :: Lens' DescribeRecoveryPointResponse (Maybe Lifecycle)

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups that are transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

describeRecoveryPointResponse_backupVaultArn :: Lens' DescribeRecoveryPointResponse (Maybe Text)

An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

describeRecoveryPointResponse_sourceBackupVaultArn :: Lens' DescribeRecoveryPointResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies the source vault where the resource was originally backed up in; for example, arn:aws:backup:us-east-1:123456789012:vault:BackupVault. If the recovery is restored to the same Amazon Web Services account or Region, this value will be null.

describeRecoveryPointResponse_lastRestoreTime :: Lens' DescribeRecoveryPointResponse (Maybe UTCTime)

The date and time that a recovery point was last restored, in Unix format and Coordinated Universal Time (UTC). The value of LastRestoreTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

describeRecoveryPointResponse_resourceArn :: Lens' DescribeRecoveryPointResponse (Maybe Text)

An ARN that uniquely identifies a saved resource. The format of the ARN depends on the resource type.

describeRecoveryPointResponse_statusMessage :: Lens' DescribeRecoveryPointResponse (Maybe Text)

A status message explaining the reason for the recovery point deletion failure.

describeRecoveryPointResponse_storageClass :: Lens' DescribeRecoveryPointResponse (Maybe StorageClass)

Specifies the storage class of the recovery point. Valid values are WARM or COLD.

describeRecoveryPointResponse_recoveryPointArn :: Lens' DescribeRecoveryPointResponse (Maybe Text)

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

describeRecoveryPointResponse_encryptionKeyArn :: Lens' DescribeRecoveryPointResponse (Maybe Text)

The server-side encryption key used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

describeRecoveryPointResponse_backupSizeInBytes :: Lens' DescribeRecoveryPointResponse (Maybe Integer)

The size, in bytes, of a backup.

describeRecoveryPointResponse_creationDate :: Lens' DescribeRecoveryPointResponse (Maybe UTCTime)

The date and time that a recovery point is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

describeRecoveryPointResponse_completionDate :: Lens' DescribeRecoveryPointResponse (Maybe UTCTime)

The date and time that a job to create a recovery point is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

describeRecoveryPointResponse_backupVaultName :: Lens' DescribeRecoveryPointResponse (Maybe Text)

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Region where they are created. They consist of lowercase letters, numbers, and hyphens.

describeRecoveryPointResponse_httpStatus :: Lens' DescribeRecoveryPointResponse Int

The response's http status code.