Network.AWS.Backup.DescribeRegionSettings

Description

Returns the current service opt-in settings for the Region. If service opt-in is enabled for a service, Backup tries to protect that service's resources in this Region, when the resource is included in an on-demand backup or scheduled backup plan. Otherwise, Backup does not try to protect that service's resources in this Region.

Creating a Request

data DescribeRegionSettings

See: newDescribeRegionSettings smart constructor.

Constructors

DescribeRegionSettings'

newDescribeRegionSettings :: DescribeRegionSettings

Create a value of DescribeRegionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeRegionSettingsResponse

See: newDescribeRegionSettingsResponse smart constructor.

Constructors

DescribeRegionSettingsResponse'

Fields

resourceTypeOptInPreference :: Maybe (HashMap Text Bool)

Returns a list of all services along with the opt-in preferences in the Region.

httpStatus :: Int

The response's http status code.

newDescribeRegionSettingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeRegionSettingsResponse'

-> DescribeRegionSettingsResponse

Create a value of DescribeRegionSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceTypeOptInPreference:DescribeRegionSettingsResponse', describeRegionSettingsResponse_resourceTypeOptInPreference - Returns a list of all services along with the opt-in preferences in the Region.

$sel:httpStatus:DescribeRegionSettingsResponse', describeRegionSettingsResponse_httpStatus - The response's http status code.

Response Lenses

describeRegionSettingsResponse_resourceTypeOptInPreference :: Lens' DescribeRegionSettingsResponse (Maybe (HashMap Text Bool))

Returns a list of all services along with the opt-in preferences in the Region.

describeRegionSettingsResponse_httpStatus :: Lens' DescribeRegionSettingsResponse Int

The response's http status code.