Network.AWS.Backup.DescribeReportPlan

Description

Returns a list of all report plans for an Amazon Web Services account and Amazon Web Services Region.

Creating a Request

data DescribeReportPlan

See: newDescribeReportPlan smart constructor.

Constructors

DescribeReportPlan'

Fields

reportPlanName :: Text

The unique name of a report plan.

newDescribeReportPlan

Arguments

:: Text
$sel:reportPlanName:DescribeReportPlan'

-> DescribeReportPlan

Create a value of DescribeReportPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportPlanName:DescribeReportPlan', describeReportPlan_reportPlanName - The unique name of a report plan.

Request Lenses

describeReportPlan_reportPlanName :: Lens' DescribeReportPlan Text

The unique name of a report plan.

Destructuring the Response

data DescribeReportPlanResponse

See: newDescribeReportPlanResponse smart constructor.

Constructors

DescribeReportPlanResponse'

Fields

reportPlan :: Maybe ReportPlan

Returns details about the report plan that is specified by its name. These details include the report plan's Amazon Resource Name (ARN), description, settings, delivery channel, deployment status, creation time, and last attempted and successful run times.

httpStatus :: Int

The response's http status code.

newDescribeReportPlanResponse

Arguments

:: Int
$sel:httpStatus:DescribeReportPlanResponse'

-> DescribeReportPlanResponse

Create a value of DescribeReportPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportPlan:DescribeReportPlanResponse', describeReportPlanResponse_reportPlan - Returns details about the report plan that is specified by its name. These details include the report plan's Amazon Resource Name (ARN), description, settings, delivery channel, deployment status, creation time, and last attempted and successful run times.

$sel:httpStatus:DescribeReportPlanResponse', describeReportPlanResponse_httpStatus - The response's http status code.

Response Lenses

describeReportPlanResponse_reportPlan :: Lens' DescribeReportPlanResponse (Maybe ReportPlan)

Returns details about the report plan that is specified by its name. These details include the report plan's Amazon Resource Name (ARN), description, settings, delivery channel, deployment status, creation time, and last attempted and successful run times.

describeReportPlanResponse_httpStatus :: Lens' DescribeReportPlanResponse Int

The response's http status code.