Network.AWS.Backup.DescribeRestoreJob

Description

Returns metadata associated with a restore job that is specified by a job ID.

Creating a Request

data DescribeRestoreJob

See: newDescribeRestoreJob smart constructor.

Constructors

DescribeRestoreJob'

Fields

restoreJobId :: Text

Uniquely identifies the job that restores a recovery point.

newDescribeRestoreJob

Arguments

:: Text
$sel:restoreJobId:DescribeRestoreJob'

-> DescribeRestoreJob

Create a value of DescribeRestoreJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:restoreJobId:DescribeRestoreJob', describeRestoreJob_restoreJobId - Uniquely identifies the job that restores a recovery point.

Request Lenses

describeRestoreJob_restoreJobId :: Lens' DescribeRestoreJob Text

Uniquely identifies the job that restores a recovery point.

Destructuring the Response

data DescribeRestoreJobResponse

See: newDescribeRestoreJobResponse smart constructor.

Constructors

DescribeRestoreJobResponse'

Fields

status :: Maybe RestoreJobStatus

Status code specifying the state of the job that is initiated by Backup to restore a recovery point.

iamRoleArn :: Maybe Text

Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

expectedCompletionTimeMinutes :: Maybe Integer

The amount of time in minutes that a job restoring a recovery point is expected to take.

restoreJobId :: Maybe Text

Uniquely identifies the job that restores a recovery point.

resourceType :: Maybe Text

Returns metadata associated with a restore job listed by resource type.

percentDone :: Maybe Text

Contains an estimated percentage that is complete of a job at the time the job status was queried.

accountId :: Maybe Text

Returns the account ID that owns the restore job.

createdResourceArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a resource whose recovery point is being restored. The format of the ARN depends on the resource type of the backed-up resource.

statusMessage :: Maybe Text

A message showing the status of a job to restore a recovery point.

recoveryPointArn :: Maybe Text

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

backupSizeInBytes :: Maybe Integer

The size, in bytes, of the restored resource.

creationDate :: Maybe POSIX

The date and time that a restore job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

completionDate :: Maybe POSIX

The date and time that a job to restore a recovery point is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

httpStatus :: Int

The response's http status code.

newDescribeRestoreJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeRestoreJobResponse'

-> DescribeRestoreJobResponse

Create a value of DescribeRestoreJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeRestoreJobResponse', describeRestoreJobResponse_status - Status code specifying the state of the job that is initiated by Backup to restore a recovery point.

$sel:iamRoleArn:DescribeRestoreJobResponse', describeRestoreJobResponse_iamRoleArn - Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

$sel:expectedCompletionTimeMinutes:DescribeRestoreJobResponse', describeRestoreJobResponse_expectedCompletionTimeMinutes - The amount of time in minutes that a job restoring a recovery point is expected to take.

$sel:restoreJobId:DescribeRestoreJob', describeRestoreJobResponse_restoreJobId - Uniquely identifies the job that restores a recovery point.

$sel:resourceType:DescribeRestoreJobResponse', describeRestoreJobResponse_resourceType - Returns metadata associated with a restore job listed by resource type.

$sel:percentDone:DescribeRestoreJobResponse', describeRestoreJobResponse_percentDone - Contains an estimated percentage that is complete of a job at the time the job status was queried.

$sel:accountId:DescribeRestoreJobResponse', describeRestoreJobResponse_accountId - Returns the account ID that owns the restore job.

$sel:createdResourceArn:DescribeRestoreJobResponse', describeRestoreJobResponse_createdResourceArn - An Amazon Resource Name (ARN) that uniquely identifies a resource whose recovery point is being restored. The format of the ARN depends on the resource type of the backed-up resource.

$sel:statusMessage:DescribeRestoreJobResponse', describeRestoreJobResponse_statusMessage - A message showing the status of a job to restore a recovery point.

$sel:recoveryPointArn:DescribeRestoreJobResponse', describeRestoreJobResponse_recoveryPointArn - An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:backupSizeInBytes:DescribeRestoreJobResponse', describeRestoreJobResponse_backupSizeInBytes - The size, in bytes, of the restored resource.

$sel:creationDate:DescribeRestoreJobResponse', describeRestoreJobResponse_creationDate - The date and time that a restore job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:completionDate:DescribeRestoreJobResponse', describeRestoreJobResponse_completionDate - The date and time that a job to restore a recovery point is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:httpStatus:DescribeRestoreJobResponse', describeRestoreJobResponse_httpStatus - The response's http status code.

Response Lenses

describeRestoreJobResponse_status :: Lens' DescribeRestoreJobResponse (Maybe RestoreJobStatus)

Status code specifying the state of the job that is initiated by Backup to restore a recovery point.

describeRestoreJobResponse_iamRoleArn :: Lens' DescribeRestoreJobResponse (Maybe Text)

Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

describeRestoreJobResponse_expectedCompletionTimeMinutes :: Lens' DescribeRestoreJobResponse (Maybe Integer)

The amount of time in minutes that a job restoring a recovery point is expected to take.

describeRestoreJobResponse_restoreJobId :: Lens' DescribeRestoreJobResponse (Maybe Text)

Uniquely identifies the job that restores a recovery point.

describeRestoreJobResponse_resourceType :: Lens' DescribeRestoreJobResponse (Maybe Text)

Returns metadata associated with a restore job listed by resource type.

describeRestoreJobResponse_percentDone :: Lens' DescribeRestoreJobResponse (Maybe Text)

Contains an estimated percentage that is complete of a job at the time the job status was queried.

describeRestoreJobResponse_accountId :: Lens' DescribeRestoreJobResponse (Maybe Text)

Returns the account ID that owns the restore job.

describeRestoreJobResponse_createdResourceArn :: Lens' DescribeRestoreJobResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a resource whose recovery point is being restored. The format of the ARN depends on the resource type of the backed-up resource.

describeRestoreJobResponse_statusMessage :: Lens' DescribeRestoreJobResponse (Maybe Text)

A message showing the status of a job to restore a recovery point.

describeRestoreJobResponse_recoveryPointArn :: Lens' DescribeRestoreJobResponse (Maybe Text)

An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

describeRestoreJobResponse_backupSizeInBytes :: Lens' DescribeRestoreJobResponse (Maybe Integer)

The size, in bytes, of the restored resource.

describeRestoreJobResponse_creationDate :: Lens' DescribeRestoreJobResponse (Maybe UTCTime)

The date and time that a restore job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

describeRestoreJobResponse_completionDate :: Lens' DescribeRestoreJobResponse (Maybe UTCTime)

The date and time that a job to restore a recovery point is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

describeRestoreJobResponse_httpStatus :: Lens' DescribeRestoreJobResponse Int

The response's http status code.