Network.AWS.Backup.GetBackupPlan

Description

Returns BackupPlan details for the specified BackupPlanId. The details are the body of a backup plan in JSON format, in addition to plan metadata.

Creating a Request

data GetBackupPlan

See: newGetBackupPlan smart constructor.

Constructors

GetBackupPlan'

Fields

versionId :: Maybe Text

Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

backupPlanId :: Text

Uniquely identifies a backup plan.

newGetBackupPlan

Arguments

:: Text
$sel:backupPlanId:GetBackupPlan'

-> GetBackupPlan

Create a value of GetBackupPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:GetBackupPlan', getBackupPlan_versionId - Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

$sel:backupPlanId:GetBackupPlan', getBackupPlan_backupPlanId - Uniquely identifies a backup plan.

Request Lenses

getBackupPlan_versionId :: Lens' GetBackupPlan (Maybe Text)

Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

getBackupPlan_backupPlanId :: Lens' GetBackupPlan Text

Uniquely identifies a backup plan.

Destructuring the Response

data GetBackupPlanResponse

See: newGetBackupPlanResponse smart constructor.

Constructors

GetBackupPlanResponse'

Fields

versionId :: Maybe Text

Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

advancedBackupSettings :: Maybe [AdvancedBackupSetting]

Contains a list of BackupOptions for each resource type. The list is populated only if the advanced option is set for the backup plan.

backupPlanId :: Maybe Text

Uniquely identifies a backup plan.

creatorRequestId :: Maybe Text

A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

backupPlanArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

lastExecutionDate :: Maybe POSIX

The last time a job to back up resources was run with this backup plan. A date and time, in Unix format and Coordinated Universal Time (UTC). The value of LastExecutionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupPlan :: Maybe BackupPlan

Specifies the body of a backup plan. Includes a BackupPlanName and one or more sets of Rules.

creationDate :: Maybe POSIX

The date and time that a backup plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

deletionDate :: Maybe POSIX

The date and time that a backup plan is deleted, in Unix format and Coordinated Universal Time (UTC). The value of DeletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

httpStatus :: Int

The response's http status code.

newGetBackupPlanResponse

Arguments

:: Int
$sel:httpStatus:GetBackupPlanResponse'

-> GetBackupPlanResponse

Create a value of GetBackupPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:GetBackupPlan', getBackupPlanResponse_versionId - Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

$sel:advancedBackupSettings:GetBackupPlanResponse', getBackupPlanResponse_advancedBackupSettings - Contains a list of BackupOptions for each resource type. The list is populated only if the advanced option is set for the backup plan.

$sel:backupPlanId:GetBackupPlan', getBackupPlanResponse_backupPlanId - Uniquely identifies a backup plan.

$sel:creatorRequestId:GetBackupPlanResponse', getBackupPlanResponse_creatorRequestId - A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

$sel:backupPlanArn:GetBackupPlanResponse', getBackupPlanResponse_backupPlanArn - An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

$sel:lastExecutionDate:GetBackupPlanResponse', getBackupPlanResponse_lastExecutionDate - The last time a job to back up resources was run with this backup plan. A date and time, in Unix format and Coordinated Universal Time (UTC). The value of LastExecutionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:backupPlan:GetBackupPlanResponse', getBackupPlanResponse_backupPlan - Specifies the body of a backup plan. Includes a BackupPlanName and one or more sets of Rules.

$sel:creationDate:GetBackupPlanResponse', getBackupPlanResponse_creationDate - The date and time that a backup plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:deletionDate:GetBackupPlanResponse', getBackupPlanResponse_deletionDate - The date and time that a backup plan is deleted, in Unix format and Coordinated Universal Time (UTC). The value of DeletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:httpStatus:GetBackupPlanResponse', getBackupPlanResponse_httpStatus - The response's http status code.

Response Lenses

getBackupPlanResponse_versionId :: Lens' GetBackupPlanResponse (Maybe Text)

Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

getBackupPlanResponse_advancedBackupSettings :: Lens' GetBackupPlanResponse (Maybe [AdvancedBackupSetting])

Contains a list of BackupOptions for each resource type. The list is populated only if the advanced option is set for the backup plan.

getBackupPlanResponse_backupPlanId :: Lens' GetBackupPlanResponse (Maybe Text)

Uniquely identifies a backup plan.

getBackupPlanResponse_creatorRequestId :: Lens' GetBackupPlanResponse (Maybe Text)

A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

getBackupPlanResponse_backupPlanArn :: Lens' GetBackupPlanResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

getBackupPlanResponse_lastExecutionDate :: Lens' GetBackupPlanResponse (Maybe UTCTime)

The last time a job to back up resources was run with this backup plan. A date and time, in Unix format and Coordinated Universal Time (UTC). The value of LastExecutionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

getBackupPlanResponse_backupPlan :: Lens' GetBackupPlanResponse (Maybe BackupPlan)

Specifies the body of a backup plan. Includes a BackupPlanName and one or more sets of Rules.

getBackupPlanResponse_creationDate :: Lens' GetBackupPlanResponse (Maybe UTCTime)

The date and time that a backup plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

getBackupPlanResponse_deletionDate :: Lens' GetBackupPlanResponse (Maybe UTCTime)

The date and time that a backup plan is deleted, in Unix format and Coordinated Universal Time (UTC). The value of DeletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

getBackupPlanResponse_httpStatus :: Lens' GetBackupPlanResponse Int

The response's http status code.