Network.AWS.Backup.GetBackupSelection

Description

Returns selection metadata and a document in JSON format that specifies a list of resources that are associated with a backup plan.

Creating a Request

data GetBackupSelection

See: newGetBackupSelection smart constructor.

Constructors

GetBackupSelection'

Fields

backupPlanId :: Text

Uniquely identifies a backup plan.

selectionId :: Text

Uniquely identifies the body of a request to assign a set of resources to a backup plan.

newGetBackupSelection

Arguments

:: Text
$sel:backupPlanId:GetBackupSelection'

-> Text
$sel:selectionId:GetBackupSelection'

-> GetBackupSelection

Create a value of GetBackupSelection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupPlanId:GetBackupSelection', getBackupSelection_backupPlanId - Uniquely identifies a backup plan.

$sel:selectionId:GetBackupSelection', getBackupSelection_selectionId - Uniquely identifies the body of a request to assign a set of resources to a backup plan.

Request Lenses

getBackupSelection_backupPlanId :: Lens' GetBackupSelection Text

Uniquely identifies a backup plan.

getBackupSelection_selectionId :: Lens' GetBackupSelection Text

Uniquely identifies the body of a request to assign a set of resources to a backup plan.

Destructuring the Response

data GetBackupSelectionResponse

See: newGetBackupSelectionResponse smart constructor.

Constructors

GetBackupSelectionResponse'

Fields

selectionId :: Maybe Text

Uniquely identifies the body of a request to assign a set of resources to a backup plan.

backupPlanId :: Maybe Text

Uniquely identifies a backup plan.

creatorRequestId :: Maybe Text

A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

creationDate :: Maybe POSIX

The date and time a backup selection is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupSelection :: Maybe BackupSelection

Specifies the body of a request to assign a set of resources to a backup plan.

httpStatus :: Int

The response's http status code.

newGetBackupSelectionResponse

Arguments

:: Int
$sel:httpStatus:GetBackupSelectionResponse'

-> GetBackupSelectionResponse

Create a value of GetBackupSelectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:selectionId:GetBackupSelection', getBackupSelectionResponse_selectionId - Uniquely identifies the body of a request to assign a set of resources to a backup plan.

$sel:backupPlanId:GetBackupSelection', getBackupSelectionResponse_backupPlanId - Uniquely identifies a backup plan.

$sel:creatorRequestId:GetBackupSelectionResponse', getBackupSelectionResponse_creatorRequestId - A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

$sel:creationDate:GetBackupSelectionResponse', getBackupSelectionResponse_creationDate - The date and time a backup selection is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:backupSelection:GetBackupSelectionResponse', getBackupSelectionResponse_backupSelection - Specifies the body of a request to assign a set of resources to a backup plan.

$sel:httpStatus:GetBackupSelectionResponse', getBackupSelectionResponse_httpStatus - The response's http status code.

Response Lenses

getBackupSelectionResponse_selectionId :: Lens' GetBackupSelectionResponse (Maybe Text)

Uniquely identifies the body of a request to assign a set of resources to a backup plan.

getBackupSelectionResponse_backupPlanId :: Lens' GetBackupSelectionResponse (Maybe Text)

Uniquely identifies a backup plan.

getBackupSelectionResponse_creatorRequestId :: Lens' GetBackupSelectionResponse (Maybe Text)

A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

getBackupSelectionResponse_creationDate :: Lens' GetBackupSelectionResponse (Maybe UTCTime)

The date and time a backup selection is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

getBackupSelectionResponse_backupSelection :: Lens' GetBackupSelectionResponse (Maybe BackupSelection)

Specifies the body of a request to assign a set of resources to a backup plan.

getBackupSelectionResponse_httpStatus :: Lens' GetBackupSelectionResponse Int

The response's http status code.