Network.AWS.Backup.ListBackupJobs

Description

Returns a list of existing backup jobs for an authenticated account for the last 30 days. For a longer period of time, consider using these monitoring tools.

Creating a Request

data ListBackupJobs

See: newListBackupJobs smart constructor.

Constructors

ListBackupJobs'

Fields

byResourceArn :: Maybe Text

Returns only backup jobs that match the specified resource Amazon Resource Name (ARN).

byCreatedAfter :: Maybe POSIX

Returns only backup jobs that were created after the specified date.

byAccountId :: Maybe Text

The account ID to list the jobs from. Returns only backup jobs associated with the specified account ID.

If used from an Organizations management account, passing * returns all jobs across the organization.

byCreatedBefore :: Maybe POSIX

Returns only backup jobs that were created before the specified date.

byBackupVaultName :: Maybe Text

Returns only backup jobs that will be stored in the specified backup vault. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

byResourceType :: Maybe Text

Returns only backup jobs for the specified resources:

DynamoDB for Amazon DynamoDB

EBS for Amazon Elastic Block Store

EC2 for Amazon Elastic Compute Cloud

EFS for Amazon Elastic File System

RDS for Amazon Relational Database Service

Aurora for Amazon Aurora

Storage Gateway for Storage Gateway

nextToken :: Maybe Text

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

byState :: Maybe BackupJobState

Returns only backup jobs that are in the specified state.

maxResults :: Maybe Natural

The maximum number of items to be returned.

newListBackupJobs :: ListBackupJobs

Create a value of ListBackupJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:byResourceArn:ListBackupJobs', listBackupJobs_byResourceArn - Returns only backup jobs that match the specified resource Amazon Resource Name (ARN).

$sel:byCreatedAfter:ListBackupJobs', listBackupJobs_byCreatedAfter - Returns only backup jobs that were created after the specified date.

$sel:byAccountId:ListBackupJobs', listBackupJobs_byAccountId - The account ID to list the jobs from. Returns only backup jobs associated with the specified account ID.

If used from an Organizations management account, passing * returns all jobs across the organization.

$sel:byCreatedBefore:ListBackupJobs', listBackupJobs_byCreatedBefore - Returns only backup jobs that were created before the specified date.

$sel:byBackupVaultName:ListBackupJobs', listBackupJobs_byBackupVaultName - Returns only backup jobs that will be stored in the specified backup vault. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

$sel:byResourceType:ListBackupJobs', listBackupJobs_byResourceType - Returns only backup jobs for the specified resources:

DynamoDB for Amazon DynamoDB

EBS for Amazon Elastic Block Store

EC2 for Amazon Elastic Compute Cloud

EFS for Amazon Elastic File System

RDS for Amazon Relational Database Service

Aurora for Amazon Aurora

Storage Gateway for Storage Gateway

$sel:nextToken:ListBackupJobs', listBackupJobs_nextToken - The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

$sel:byState:ListBackupJobs', listBackupJobs_byState - Returns only backup jobs that are in the specified state.

$sel:maxResults:ListBackupJobs', listBackupJobs_maxResults - The maximum number of items to be returned.

Request Lenses

listBackupJobs_byResourceArn :: Lens' ListBackupJobs (Maybe Text)

Returns only backup jobs that match the specified resource Amazon Resource Name (ARN).

listBackupJobs_byCreatedAfter :: Lens' ListBackupJobs (Maybe UTCTime)

Returns only backup jobs that were created after the specified date.

listBackupJobs_byAccountId :: Lens' ListBackupJobs (Maybe Text)

The account ID to list the jobs from. Returns only backup jobs associated with the specified account ID.

If used from an Organizations management account, passing * returns all jobs across the organization.

listBackupJobs_byCreatedBefore :: Lens' ListBackupJobs (Maybe UTCTime)

Returns only backup jobs that were created before the specified date.

listBackupJobs_byBackupVaultName :: Lens' ListBackupJobs (Maybe Text)

Returns only backup jobs that will be stored in the specified backup vault. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

listBackupJobs_byResourceType :: Lens' ListBackupJobs (Maybe Text)

Returns only backup jobs for the specified resources:

DynamoDB for Amazon DynamoDB

EBS for Amazon Elastic Block Store

EC2 for Amazon Elastic Compute Cloud

EFS for Amazon Elastic File System

RDS for Amazon Relational Database Service

Aurora for Amazon Aurora

Storage Gateway for Storage Gateway

listBackupJobs_nextToken :: Lens' ListBackupJobs (Maybe Text)

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

listBackupJobs_byState :: Lens' ListBackupJobs (Maybe BackupJobState)

Returns only backup jobs that are in the specified state.

listBackupJobs_maxResults :: Lens' ListBackupJobs (Maybe Natural)

The maximum number of items to be returned.

Destructuring the Response

data ListBackupJobsResponse

See: newListBackupJobsResponse smart constructor.

Constructors

ListBackupJobsResponse'

Fields

backupJobs :: Maybe [BackupJob]

An array of structures containing metadata about your backup jobs returned in JSON format.

nextToken :: Maybe Text

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

httpStatus :: Int

The response's http status code.

newListBackupJobsResponse

Arguments

:: Int
$sel:httpStatus:ListBackupJobsResponse'

-> ListBackupJobsResponse

Create a value of ListBackupJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupJobs:ListBackupJobsResponse', listBackupJobsResponse_backupJobs - An array of structures containing metadata about your backup jobs returned in JSON format.

$sel:nextToken:ListBackupJobs', listBackupJobsResponse_nextToken - The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

$sel:httpStatus:ListBackupJobsResponse', listBackupJobsResponse_httpStatus - The response's http status code.

Response Lenses

listBackupJobsResponse_backupJobs :: Lens' ListBackupJobsResponse (Maybe [BackupJob])

An array of structures containing metadata about your backup jobs returned in JSON format.

listBackupJobsResponse_nextToken :: Lens' ListBackupJobsResponse (Maybe Text)

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

listBackupJobsResponse_httpStatus :: Lens' ListBackupJobsResponse Int

The response's http status code.