Network.AWS.Backup.ListFrameworks

Description

Returns a list of all frameworks for an Amazon Web Services account and Amazon Web Services Region.

Creating a Request

data ListFrameworks

See: newListFrameworks smart constructor.

Constructors

ListFrameworks'

Fields

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

maxResults :: Maybe Natural

The number of desired results from 1 to 1000. Optional. If unspecified, the query will return 1 MB of data.

newListFrameworks :: ListFrameworks

Create a value of ListFrameworks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFrameworks', listFrameworks_nextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

$sel:maxResults:ListFrameworks', listFrameworks_maxResults - The number of desired results from 1 to 1000. Optional. If unspecified, the query will return 1 MB of data.

Request Lenses

listFrameworks_nextToken :: Lens' ListFrameworks (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listFrameworks_maxResults :: Lens' ListFrameworks (Maybe Natural)

The number of desired results from 1 to 1000. Optional. If unspecified, the query will return 1 MB of data.

Destructuring the Response

data ListFrameworksResponse

See: newListFrameworksResponse smart constructor.

Constructors

ListFrameworksResponse'

Fields

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

frameworks :: Maybe [Framework]

A list of frameworks with details for each framework, including the framework name, Amazon Resource Name (ARN), description, number of controls, creation time, and deployment status.

httpStatus :: Int

The response's http status code.

newListFrameworksResponse

Arguments

:: Int
$sel:httpStatus:ListFrameworksResponse'

-> ListFrameworksResponse

Create a value of ListFrameworksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFrameworks', listFrameworksResponse_nextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

$sel:frameworks:ListFrameworksResponse', listFrameworksResponse_frameworks - A list of frameworks with details for each framework, including the framework name, Amazon Resource Name (ARN), description, number of controls, creation time, and deployment status.

$sel:httpStatus:ListFrameworksResponse', listFrameworksResponse_httpStatus - The response's http status code.

Response Lenses

listFrameworksResponse_nextToken :: Lens' ListFrameworksResponse (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listFrameworksResponse_frameworks :: Lens' ListFrameworksResponse (Maybe [Framework])

A list of frameworks with details for each framework, including the framework name, Amazon Resource Name (ARN), description, number of controls, creation time, and deployment status.

listFrameworksResponse_httpStatus :: Lens' ListFrameworksResponse Int

The response's http status code.