Network.AWS.Backup.ListRecoveryPointsByBackupVault

Description

Returns detailed information about the recovery points stored in a backup vault.

Creating a Request

data ListRecoveryPointsByBackupVault

See: newListRecoveryPointsByBackupVault smart constructor.

Constructors

ListRecoveryPointsByBackupVault'

Fields

byResourceArn :: Maybe Text

Returns only recovery points that match the specified resource Amazon Resource Name (ARN).

byCreatedAfter :: Maybe POSIX

Returns only recovery points that were created after the specified timestamp.

byCreatedBefore :: Maybe POSIX

Returns only recovery points that were created before the specified timestamp.

byBackupPlanId :: Maybe Text

Returns only recovery points that match the specified backup plan ID.

byResourceType :: Maybe Text

Returns only recovery points that match the specified resource type.

nextToken :: Maybe Text

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

maxResults :: Maybe Natural

The maximum number of items to be returned.

backupVaultName :: Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Backup vault name might not be available when a supported service creates the backup.

newListRecoveryPointsByBackupVault

Arguments

:: Text
$sel:backupVaultName:ListRecoveryPointsByBackupVault'

-> ListRecoveryPointsByBackupVault

Create a value of ListRecoveryPointsByBackupVault with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:byResourceArn:ListRecoveryPointsByBackupVault', listRecoveryPointsByBackupVault_byResourceArn - Returns only recovery points that match the specified resource Amazon Resource Name (ARN).

$sel:byCreatedAfter:ListRecoveryPointsByBackupVault', listRecoveryPointsByBackupVault_byCreatedAfter - Returns only recovery points that were created after the specified timestamp.

$sel:byCreatedBefore:ListRecoveryPointsByBackupVault', listRecoveryPointsByBackupVault_byCreatedBefore - Returns only recovery points that were created before the specified timestamp.

$sel:byBackupPlanId:ListRecoveryPointsByBackupVault', listRecoveryPointsByBackupVault_byBackupPlanId - Returns only recovery points that match the specified backup plan ID.

$sel:byResourceType:ListRecoveryPointsByBackupVault', listRecoveryPointsByBackupVault_byResourceType - Returns only recovery points that match the specified resource type.

$sel:nextToken:ListRecoveryPointsByBackupVault', listRecoveryPointsByBackupVault_nextToken - The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

$sel:maxResults:ListRecoveryPointsByBackupVault', listRecoveryPointsByBackupVault_maxResults - The maximum number of items to be returned.

$sel:backupVaultName:ListRecoveryPointsByBackupVault', listRecoveryPointsByBackupVault_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Backup vault name might not be available when a supported service creates the backup.

Request Lenses

listRecoveryPointsByBackupVault_byResourceArn :: Lens' ListRecoveryPointsByBackupVault (Maybe Text)

Returns only recovery points that match the specified resource Amazon Resource Name (ARN).

listRecoveryPointsByBackupVault_byCreatedAfter :: Lens' ListRecoveryPointsByBackupVault (Maybe UTCTime)

Returns only recovery points that were created after the specified timestamp.

listRecoveryPointsByBackupVault_byCreatedBefore :: Lens' ListRecoveryPointsByBackupVault (Maybe UTCTime)

Returns only recovery points that were created before the specified timestamp.

listRecoveryPointsByBackupVault_byBackupPlanId :: Lens' ListRecoveryPointsByBackupVault (Maybe Text)

Returns only recovery points that match the specified backup plan ID.

listRecoveryPointsByBackupVault_byResourceType :: Lens' ListRecoveryPointsByBackupVault (Maybe Text)

Returns only recovery points that match the specified resource type.

listRecoveryPointsByBackupVault_nextToken :: Lens' ListRecoveryPointsByBackupVault (Maybe Text)

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

listRecoveryPointsByBackupVault_maxResults :: Lens' ListRecoveryPointsByBackupVault (Maybe Natural)

The maximum number of items to be returned.

listRecoveryPointsByBackupVault_backupVaultName :: Lens' ListRecoveryPointsByBackupVault Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Backup vault name might not be available when a supported service creates the backup.

Destructuring the Response

data ListRecoveryPointsByBackupVaultResponse

See: newListRecoveryPointsByBackupVaultResponse smart constructor.

Constructors

ListRecoveryPointsByBackupVaultResponse'

Fields

recoveryPoints :: Maybe [RecoveryPointByBackupVault]

An array of objects that contain detailed information about recovery points saved in a backup vault.

nextToken :: Maybe Text

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

httpStatus :: Int

The response's http status code.

newListRecoveryPointsByBackupVaultResponse

Arguments

:: Int
$sel:httpStatus:ListRecoveryPointsByBackupVaultResponse'

-> ListRecoveryPointsByBackupVaultResponse

Create a value of ListRecoveryPointsByBackupVaultResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recoveryPoints:ListRecoveryPointsByBackupVaultResponse', listRecoveryPointsByBackupVaultResponse_recoveryPoints - An array of objects that contain detailed information about recovery points saved in a backup vault.

$sel:nextToken:ListRecoveryPointsByBackupVault', listRecoveryPointsByBackupVaultResponse_nextToken - The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

$sel:httpStatus:ListRecoveryPointsByBackupVaultResponse', listRecoveryPointsByBackupVaultResponse_httpStatus - The response's http status code.

Response Lenses

listRecoveryPointsByBackupVaultResponse_recoveryPoints :: Lens' ListRecoveryPointsByBackupVaultResponse (Maybe [RecoveryPointByBackupVault])

An array of objects that contain detailed information about recovery points saved in a backup vault.

listRecoveryPointsByBackupVaultResponse_nextToken :: Lens' ListRecoveryPointsByBackupVaultResponse (Maybe Text)

The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

listRecoveryPointsByBackupVaultResponse_httpStatus :: Lens' ListRecoveryPointsByBackupVaultResponse Int

The response's http status code.