Network.AWS.Backup.ListReportJobs

Description

Returns details about your report jobs.

Creating a Request

data ListReportJobs

See: newListReportJobs smart constructor.

Constructors

ListReportJobs'

Fields

byStatus :: Maybe Text

Returns only report jobs that are in the specified status. The statuses are:

CREATED | RUNNING | COMPLETED | FAILED

byReportPlanName :: Maybe Text

Returns only report jobs with the specified report plan name.

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

byCreationBefore :: Maybe POSIX

Returns only report jobs that were created before the date and time specified in Unix format and Coordinated Universal Time (UTC). For example, the value 1516925490 represents Friday, January 26, 2018 12:11:30 AM.

byCreationAfter :: Maybe POSIX

Returns only report jobs that were created after the date and time specified in Unix format and Coordinated Universal Time (UTC). For example, the value 1516925490 represents Friday, January 26, 2018 12:11:30 AM.

maxResults :: Maybe Natural

The number of desired results from 1 to 1000. Optional. If unspecified, the query will return 1 MB of data.

newListReportJobs :: ListReportJobs

Create a value of ListReportJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:byStatus:ListReportJobs', listReportJobs_byStatus - Returns only report jobs that are in the specified status. The statuses are:

CREATED | RUNNING | COMPLETED | FAILED

$sel:byReportPlanName:ListReportJobs', listReportJobs_byReportPlanName - Returns only report jobs with the specified report plan name.

$sel:nextToken:ListReportJobs', listReportJobs_nextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

$sel:byCreationBefore:ListReportJobs', listReportJobs_byCreationBefore - Returns only report jobs that were created before the date and time specified in Unix format and Coordinated Universal Time (UTC). For example, the value 1516925490 represents Friday, January 26, 2018 12:11:30 AM.

$sel:byCreationAfter:ListReportJobs', listReportJobs_byCreationAfter - Returns only report jobs that were created after the date and time specified in Unix format and Coordinated Universal Time (UTC). For example, the value 1516925490 represents Friday, January 26, 2018 12:11:30 AM.

$sel:maxResults:ListReportJobs', listReportJobs_maxResults - The number of desired results from 1 to 1000. Optional. If unspecified, the query will return 1 MB of data.

Request Lenses

listReportJobs_byStatus :: Lens' ListReportJobs (Maybe Text)

Returns only report jobs that are in the specified status. The statuses are:

CREATED | RUNNING | COMPLETED | FAILED

listReportJobs_byReportPlanName :: Lens' ListReportJobs (Maybe Text)

Returns only report jobs with the specified report plan name.

listReportJobs_nextToken :: Lens' ListReportJobs (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listReportJobs_byCreationBefore :: Lens' ListReportJobs (Maybe UTCTime)

Returns only report jobs that were created before the date and time specified in Unix format and Coordinated Universal Time (UTC). For example, the value 1516925490 represents Friday, January 26, 2018 12:11:30 AM.

listReportJobs_byCreationAfter :: Lens' ListReportJobs (Maybe UTCTime)

Returns only report jobs that were created after the date and time specified in Unix format and Coordinated Universal Time (UTC). For example, the value 1516925490 represents Friday, January 26, 2018 12:11:30 AM.

listReportJobs_maxResults :: Lens' ListReportJobs (Maybe Natural)

The number of desired results from 1 to 1000. Optional. If unspecified, the query will return 1 MB of data.

Destructuring the Response

data ListReportJobsResponse

See: newListReportJobsResponse smart constructor.

Constructors

ListReportJobsResponse'

Fields

reportJobs :: Maybe [ReportJob]

Details about your report jobs in JSON format.

nextToken :: Maybe Text

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

httpStatus :: Int

The response's http status code.

newListReportJobsResponse

Arguments

:: Int
$sel:httpStatus:ListReportJobsResponse'

-> ListReportJobsResponse

Create a value of ListReportJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportJobs:ListReportJobsResponse', listReportJobsResponse_reportJobs - Details about your report jobs in JSON format.

$sel:nextToken:ListReportJobs', listReportJobsResponse_nextToken - An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

$sel:httpStatus:ListReportJobsResponse', listReportJobsResponse_httpStatus - The response's http status code.

Response Lenses

listReportJobsResponse_reportJobs :: Lens' ListReportJobsResponse (Maybe [ReportJob])

Details about your report jobs in JSON format.

listReportJobsResponse_nextToken :: Lens' ListReportJobsResponse (Maybe Text)

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

listReportJobsResponse_httpStatus :: Lens' ListReportJobsResponse Int

The response's http status code.