Network.AWS.Backup.PutBackupVaultAccessPolicy

Description

Sets a resource-based policy that is used to manage access permissions on the target backup vault. Requires a backup vault name and an access policy document in JSON format.

Creating a Request

data PutBackupVaultAccessPolicy

See: newPutBackupVaultAccessPolicy smart constructor.

Constructors

PutBackupVaultAccessPolicy'

Fields

policy :: Maybe Text

The backup vault access policy document in JSON format.

backupVaultName :: Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

newPutBackupVaultAccessPolicy

Arguments

:: Text
$sel:backupVaultName:PutBackupVaultAccessPolicy'

-> PutBackupVaultAccessPolicy

Create a value of PutBackupVaultAccessPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:PutBackupVaultAccessPolicy', putBackupVaultAccessPolicy_policy - The backup vault access policy document in JSON format.

$sel:backupVaultName:PutBackupVaultAccessPolicy', putBackupVaultAccessPolicy_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Request Lenses

putBackupVaultAccessPolicy_policy :: Lens' PutBackupVaultAccessPolicy (Maybe Text)

The backup vault access policy document in JSON format.

putBackupVaultAccessPolicy_backupVaultName :: Lens' PutBackupVaultAccessPolicy Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Destructuring the Response

data PutBackupVaultAccessPolicyResponse

See: newPutBackupVaultAccessPolicyResponse smart constructor.

Constructors

PutBackupVaultAccessPolicyResponse'

newPutBackupVaultAccessPolicyResponse :: PutBackupVaultAccessPolicyResponse

Create a value of PutBackupVaultAccessPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.