Network.AWS.Backup.StartCopyJob

Description

Starts a job to create a one-time copy of the specified resource.

Does not support continuous backups.

Creating a Request

data StartCopyJob

See: newStartCopyJob smart constructor.

Constructors

StartCopyJob'

Fields

idempotencyToken :: Maybe Text

A customer-chosen string that you can use to distinguish between otherwise identical calls to StartCopyJob. Retrying a successful request with the same idempotency token results in a success message with no action taken.

lifecycle :: Maybe Lifecycle

recoveryPointArn :: Text

An ARN that uniquely identifies a recovery point to use for the copy job; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

sourceBackupVaultName :: Text

The name of a logical source container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

destinationBackupVaultArn :: Text

An Amazon Resource Name (ARN) that uniquely identifies a destination backup vault to copy to; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

iamRoleArn :: Text

Specifies the IAM role ARN used to copy the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

newStartCopyJob

Arguments

:: Text
$sel:recoveryPointArn:StartCopyJob'

-> Text
$sel:sourceBackupVaultName:StartCopyJob'

-> Text
$sel:destinationBackupVaultArn:StartCopyJob'

-> Text
$sel:iamRoleArn:StartCopyJob'

-> StartCopyJob

Create a value of StartCopyJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idempotencyToken:StartCopyJob', startCopyJob_idempotencyToken - A customer-chosen string that you can use to distinguish between otherwise identical calls to StartCopyJob. Retrying a successful request with the same idempotency token results in a success message with no action taken.

$sel:lifecycle:StartCopyJob', startCopyJob_lifecycle - Undocumented member.

$sel:recoveryPointArn:StartCopyJob', startCopyJob_recoveryPointArn - An ARN that uniquely identifies a recovery point to use for the copy job; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:sourceBackupVaultName:StartCopyJob', startCopyJob_sourceBackupVaultName - The name of a logical source container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

$sel:destinationBackupVaultArn:StartCopyJob', startCopyJob_destinationBackupVaultArn - An Amazon Resource Name (ARN) that uniquely identifies a destination backup vault to copy to; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:iamRoleArn:StartCopyJob', startCopyJob_iamRoleArn - Specifies the IAM role ARN used to copy the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

Request Lenses

startCopyJob_idempotencyToken :: Lens' StartCopyJob (Maybe Text)

A customer-chosen string that you can use to distinguish between otherwise identical calls to StartCopyJob. Retrying a successful request with the same idempotency token results in a success message with no action taken.

startCopyJob_lifecycle :: Lens' StartCopyJob (Maybe Lifecycle)

Undocumented member.

startCopyJob_recoveryPointArn :: Lens' StartCopyJob Text

An ARN that uniquely identifies a recovery point to use for the copy job; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

startCopyJob_sourceBackupVaultName :: Lens' StartCopyJob Text

The name of a logical source container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

startCopyJob_destinationBackupVaultArn :: Lens' StartCopyJob Text

An Amazon Resource Name (ARN) that uniquely identifies a destination backup vault to copy to; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

startCopyJob_iamRoleArn :: Lens' StartCopyJob Text

Specifies the IAM role ARN used to copy the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

Destructuring the Response

data StartCopyJobResponse

See: newStartCopyJobResponse smart constructor.

Constructors

StartCopyJobResponse'

Fields

copyJobId :: Maybe Text

Uniquely identifies a copy job.

creationDate :: Maybe POSIX

The date and time that a copy job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

httpStatus :: Int

The response's http status code.

newStartCopyJobResponse

Arguments

:: Int
$sel:httpStatus:StartCopyJobResponse'

-> StartCopyJobResponse

Create a value of StartCopyJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:copyJobId:StartCopyJobResponse', startCopyJobResponse_copyJobId - Uniquely identifies a copy job.

$sel:creationDate:StartCopyJobResponse', startCopyJobResponse_creationDate - The date and time that a copy job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:httpStatus:StartCopyJobResponse', startCopyJobResponse_httpStatus - The response's http status code.

Response Lenses

startCopyJobResponse_copyJobId :: Lens' StartCopyJobResponse (Maybe Text)

Uniquely identifies a copy job.

startCopyJobResponse_creationDate :: Lens' StartCopyJobResponse (Maybe UTCTime)

The date and time that a copy job is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

startCopyJobResponse_httpStatus :: Lens' StartCopyJobResponse Int

The response's http status code.