Network.AWS.Backup.Types.BackupRule

Description

Documentation

data BackupRule

Specifies a scheduled task used to back up a selection of resources.

See: newBackupRule smart constructor.

Constructors

BackupRule'

Fields

ruleId :: Maybe Text

Uniquely identifies a rule that is used to schedule the backup of a selection of resources.

lifecycle :: Maybe Lifecycle

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

recoveryPointTags :: Maybe (Sensitive (HashMap Text Text))

An array of key-value pair strings that are assigned to resources that are associated with this rule when restored from backup.

scheduleExpression :: Maybe Text

A cron expression in UTC specifying when Backup initiates a backup job. For more information about cron expressions, see Schedule Expressions for Rules in the Amazon CloudWatch Events User Guide.. Prior to specifying a value for this parameter, we recommend testing your cron expression using one of the many available cron generator and testing tools.

enableContinuousBackup :: Maybe Bool

Specifies whether Backup creates continuous backups. True causes Backup to create continuous backups capable of point-in-time restore (PITR). False (or not specified) causes Backup to create snapshot backups.

completionWindowMinutes :: Maybe Integer

A value in minutes after a backup job is successfully started before it must be completed or it will be canceled by Backup. This value is optional.

copyActions :: Maybe [CopyAction]

An array of CopyAction objects, which contains the details of the copy operation.

startWindowMinutes :: Maybe Integer

A value in minutes after a backup is scheduled before a job will be canceled if it doesn't start successfully. This value is optional.

ruleName :: Text

An optional display name for a backup rule.

targetBackupVaultName :: Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

newBackupRule

Arguments

:: Text
$sel:ruleName:BackupRule'

-> Text
$sel:targetBackupVaultName:BackupRule'

-> BackupRule

Create a value of BackupRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleId:BackupRule', backupRule_ruleId - Uniquely identifies a rule that is used to schedule the backup of a selection of resources.

$sel:lifecycle:BackupRule', backupRule_lifecycle - The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

$sel:recoveryPointTags:BackupRule', backupRule_recoveryPointTags - An array of key-value pair strings that are assigned to resources that are associated with this rule when restored from backup.

$sel:scheduleExpression:BackupRule', backupRule_scheduleExpression - A cron expression in UTC specifying when Backup initiates a backup job. For more information about cron expressions, see Schedule Expressions for Rules in the Amazon CloudWatch Events User Guide.. Prior to specifying a value for this parameter, we recommend testing your cron expression using one of the many available cron generator and testing tools.

$sel:enableContinuousBackup:BackupRule', backupRule_enableContinuousBackup - Specifies whether Backup creates continuous backups. True causes Backup to create continuous backups capable of point-in-time restore (PITR). False (or not specified) causes Backup to create snapshot backups.

$sel:completionWindowMinutes:BackupRule', backupRule_completionWindowMinutes - A value in minutes after a backup job is successfully started before it must be completed or it will be canceled by Backup. This value is optional.

$sel:copyActions:BackupRule', backupRule_copyActions - An array of CopyAction objects, which contains the details of the copy operation.

$sel:startWindowMinutes:BackupRule', backupRule_startWindowMinutes - A value in minutes after a backup is scheduled before a job will be canceled if it doesn't start successfully. This value is optional.

$sel:ruleName:BackupRule', backupRule_ruleName - An optional display name for a backup rule.

$sel:targetBackupVaultName:BackupRule', backupRule_targetBackupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

backupRule_ruleId :: Lens' BackupRule (Maybe Text)

Uniquely identifies a rule that is used to schedule the backup of a selection of resources.

backupRule_lifecycle :: Lens' BackupRule (Maybe Lifecycle)

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

backupRule_recoveryPointTags :: Lens' BackupRule (Maybe (HashMap Text Text))

An array of key-value pair strings that are assigned to resources that are associated with this rule when restored from backup.

backupRule_scheduleExpression :: Lens' BackupRule (Maybe Text)

A cron expression in UTC specifying when Backup initiates a backup job. For more information about cron expressions, see Schedule Expressions for Rules in the Amazon CloudWatch Events User Guide.. Prior to specifying a value for this parameter, we recommend testing your cron expression using one of the many available cron generator and testing tools.

backupRule_enableContinuousBackup :: Lens' BackupRule (Maybe Bool)

Specifies whether Backup creates continuous backups. True causes Backup to create continuous backups capable of point-in-time restore (PITR). False (or not specified) causes Backup to create snapshot backups.

backupRule_completionWindowMinutes :: Lens' BackupRule (Maybe Integer)

A value in minutes after a backup job is successfully started before it must be completed or it will be canceled by Backup. This value is optional.

backupRule_copyActions :: Lens' BackupRule (Maybe [CopyAction])

An array of CopyAction objects, which contains the details of the copy operation.

backupRule_startWindowMinutes :: Lens' BackupRule (Maybe Integer)

A value in minutes after a backup is scheduled before a job will be canceled if it doesn't start successfully. This value is optional.

backupRule_ruleName :: Lens' BackupRule Text

An optional display name for a backup rule.

backupRule_targetBackupVaultName :: Lens' BackupRule Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.