Network.AWS.Backup.Types.BackupRuleInput

Description

Documentation

data BackupRuleInput

Specifies a scheduled task used to back up a selection of resources.

See: newBackupRuleInput smart constructor.

Constructors

BackupRuleInput'

Fields

lifecycle :: Maybe Lifecycle

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup will transition and expire backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

recoveryPointTags :: Maybe (Sensitive (HashMap Text Text))

To help organize your resources, you can assign your own metadata to the resources that you create. Each tag is a key-value pair.

scheduleExpression :: Maybe Text

A CRON expression in UTC specifying when Backup initiates a backup job.

enableContinuousBackup :: Maybe Bool

Specifies whether Backup creates continuous backups. True causes Backup to create continuous backups capable of point-in-time restore (PITR). False (or not specified) causes Backup to create snapshot backups.

completionWindowMinutes :: Maybe Integer

A value in minutes after a backup job is successfully started before it must be completed or it will be canceled by Backup. This value is optional.

copyActions :: Maybe [CopyAction]

An array of CopyAction objects, which contains the details of the copy operation.

startWindowMinutes :: Maybe Integer

A value in minutes after a backup is scheduled before a job will be canceled if it doesn't start successfully. This value is optional.

ruleName :: Text

An optional display name for a backup rule.

targetBackupVaultName :: Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

newBackupRuleInput

Arguments

:: Text
$sel:ruleName:BackupRuleInput'

-> Text
$sel:targetBackupVaultName:BackupRuleInput'

-> BackupRuleInput

Create a value of BackupRuleInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lifecycle:BackupRuleInput', backupRuleInput_lifecycle - The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup will transition and expire backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

$sel:recoveryPointTags:BackupRuleInput', backupRuleInput_recoveryPointTags - To help organize your resources, you can assign your own metadata to the resources that you create. Each tag is a key-value pair.

$sel:scheduleExpression:BackupRuleInput', backupRuleInput_scheduleExpression - A CRON expression in UTC specifying when Backup initiates a backup job.

$sel:enableContinuousBackup:BackupRuleInput', backupRuleInput_enableContinuousBackup - Specifies whether Backup creates continuous backups. True causes Backup to create continuous backups capable of point-in-time restore (PITR). False (or not specified) causes Backup to create snapshot backups.

$sel:completionWindowMinutes:BackupRuleInput', backupRuleInput_completionWindowMinutes - A value in minutes after a backup job is successfully started before it must be completed or it will be canceled by Backup. This value is optional.

$sel:copyActions:BackupRuleInput', backupRuleInput_copyActions - An array of CopyAction objects, which contains the details of the copy operation.

$sel:startWindowMinutes:BackupRuleInput', backupRuleInput_startWindowMinutes - A value in minutes after a backup is scheduled before a job will be canceled if it doesn't start successfully. This value is optional.

$sel:ruleName:BackupRuleInput', backupRuleInput_ruleName - An optional display name for a backup rule.

$sel:targetBackupVaultName:BackupRuleInput', backupRuleInput_targetBackupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

backupRuleInput_lifecycle :: Lens' BackupRuleInput (Maybe Lifecycle)

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup will transition and expire backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

backupRuleInput_recoveryPointTags :: Lens' BackupRuleInput (Maybe (HashMap Text Text))

To help organize your resources, you can assign your own metadata to the resources that you create. Each tag is a key-value pair.

backupRuleInput_scheduleExpression :: Lens' BackupRuleInput (Maybe Text)

A CRON expression in UTC specifying when Backup initiates a backup job.

backupRuleInput_enableContinuousBackup :: Lens' BackupRuleInput (Maybe Bool)

Specifies whether Backup creates continuous backups. True causes Backup to create continuous backups capable of point-in-time restore (PITR). False (or not specified) causes Backup to create snapshot backups.

backupRuleInput_completionWindowMinutes :: Lens' BackupRuleInput (Maybe Integer)

A value in minutes after a backup job is successfully started before it must be completed or it will be canceled by Backup. This value is optional.

backupRuleInput_copyActions :: Lens' BackupRuleInput (Maybe [CopyAction])

An array of CopyAction objects, which contains the details of the copy operation.

backupRuleInput_startWindowMinutes :: Lens' BackupRuleInput (Maybe Integer)

A value in minutes after a backup is scheduled before a job will be canceled if it doesn't start successfully. This value is optional.

backupRuleInput_ruleName :: Lens' BackupRuleInput Text

An optional display name for a backup rule.

backupRuleInput_targetBackupVaultName :: Lens' BackupRuleInput Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.