Network.AWS.Backup.Types.BackupVaultListMember

Description

Documentation

data BackupVaultListMember

Contains metadata about a backup vault.

See: newBackupVaultListMember smart constructor.

Constructors

BackupVaultListMember'

Fields

lockDate :: Maybe POSIX

The date and time when Backup Vault Lock configuration becomes immutable, meaning it cannot be changed or deleted.

If you applied Vault Lock to your vault without specifying a lock date, you can change your Vault Lock settings, or delete Vault Lock from the vault entirely, at any time.

This value is in Unix format, Coordinated Universal Time (UTC), and accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

maxRetentionDays :: Maybe Integer

The Backup Vault Lock setting that specifies the maximum retention period that the vault retains its recovery points. If this parameter is not specified, Vault Lock does not enforce a maximum retention period on the recovery points in the vault (allowing indefinite storage).

If specified, any backup or copy job to the vault must have a lifecycle policy with a retention period equal to or shorter than the maximum retention period. If the job's retention period is longer than that maximum retention period, then the vault fails the backup or copy job, and you should either modify your lifecycle settings or use a different vault. Recovery points already stored in the vault prior to Vault Lock are not affected.

locked :: Maybe Bool

A Boolean value that indicates whether Backup Vault Lock applies to the selected backup vault. If true, Vault Lock prevents delete and update operations on the recovery points in the selected vault.

creatorRequestId :: Maybe Text

A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

numberOfRecoveryPoints :: Maybe Integer

The number of recovery points that are stored in a backup vault.

backupVaultArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

encryptionKeyArn :: Maybe Text

The server-side encryption key that is used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

creationDate :: Maybe POSIX

The date and time a resource backup is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupVaultName :: Maybe Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

minRetentionDays :: Maybe Integer

The Backup Vault Lock setting that specifies the minimum retention period that the vault retains its recovery points. If this parameter is not specified, Vault Lock does not enforce a minimum retention period.

If specified, any backup or copy job to the vault must have a lifecycle policy with a retention period equal to or longer than the minimum retention period. If the job's retention period is shorter than that minimum retention period, then the vault fails the backup or copy job, and you should either modify your lifecycle settings or use a different vault. Recovery points already stored in the vault prior to Vault Lock are not affected.

newBackupVaultListMember :: BackupVaultListMember

Create a value of BackupVaultListMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lockDate:BackupVaultListMember', backupVaultListMember_lockDate - The date and time when Backup Vault Lock configuration becomes immutable, meaning it cannot be changed or deleted.

If you applied Vault Lock to your vault without specifying a lock date, you can change your Vault Lock settings, or delete Vault Lock from the vault entirely, at any time.

This value is in Unix format, Coordinated Universal Time (UTC), and accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:maxRetentionDays:BackupVaultListMember', backupVaultListMember_maxRetentionDays - The Backup Vault Lock setting that specifies the maximum retention period that the vault retains its recovery points. If this parameter is not specified, Vault Lock does not enforce a maximum retention period on the recovery points in the vault (allowing indefinite storage).

If specified, any backup or copy job to the vault must have a lifecycle policy with a retention period equal to or shorter than the maximum retention period. If the job's retention period is longer than that maximum retention period, then the vault fails the backup or copy job, and you should either modify your lifecycle settings or use a different vault. Recovery points already stored in the vault prior to Vault Lock are not affected.

$sel:locked:BackupVaultListMember', backupVaultListMember_locked - A Boolean value that indicates whether Backup Vault Lock applies to the selected backup vault. If true, Vault Lock prevents delete and update operations on the recovery points in the selected vault.

$sel:creatorRequestId:BackupVaultListMember', backupVaultListMember_creatorRequestId - A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

$sel:numberOfRecoveryPoints:BackupVaultListMember', backupVaultListMember_numberOfRecoveryPoints - The number of recovery points that are stored in a backup vault.

$sel:backupVaultArn:BackupVaultListMember', backupVaultListMember_backupVaultArn - An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:encryptionKeyArn:BackupVaultListMember', backupVaultListMember_encryptionKeyArn - The server-side encryption key that is used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

$sel:creationDate:BackupVaultListMember', backupVaultListMember_creationDate - The date and time a resource backup is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:backupVaultName:BackupVaultListMember', backupVaultListMember_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

$sel:minRetentionDays:BackupVaultListMember', backupVaultListMember_minRetentionDays - The Backup Vault Lock setting that specifies the minimum retention period that the vault retains its recovery points. If this parameter is not specified, Vault Lock does not enforce a minimum retention period.

If specified, any backup or copy job to the vault must have a lifecycle policy with a retention period equal to or longer than the minimum retention period. If the job's retention period is shorter than that minimum retention period, then the vault fails the backup or copy job, and you should either modify your lifecycle settings or use a different vault. Recovery points already stored in the vault prior to Vault Lock are not affected.

backupVaultListMember_lockDate :: Lens' BackupVaultListMember (Maybe UTCTime)

The date and time when Backup Vault Lock configuration becomes immutable, meaning it cannot be changed or deleted.

If you applied Vault Lock to your vault without specifying a lock date, you can change your Vault Lock settings, or delete Vault Lock from the vault entirely, at any time.

This value is in Unix format, Coordinated Universal Time (UTC), and accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupVaultListMember_maxRetentionDays :: Lens' BackupVaultListMember (Maybe Integer)

The Backup Vault Lock setting that specifies the maximum retention period that the vault retains its recovery points. If this parameter is not specified, Vault Lock does not enforce a maximum retention period on the recovery points in the vault (allowing indefinite storage).

If specified, any backup or copy job to the vault must have a lifecycle policy with a retention period equal to or shorter than the maximum retention period. If the job's retention period is longer than that maximum retention period, then the vault fails the backup or copy job, and you should either modify your lifecycle settings or use a different vault. Recovery points already stored in the vault prior to Vault Lock are not affected.

backupVaultListMember_locked :: Lens' BackupVaultListMember (Maybe Bool)

A Boolean value that indicates whether Backup Vault Lock applies to the selected backup vault. If true, Vault Lock prevents delete and update operations on the recovery points in the selected vault.

backupVaultListMember_creatorRequestId :: Lens' BackupVaultListMember (Maybe Text)

A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

backupVaultListMember_numberOfRecoveryPoints :: Lens' BackupVaultListMember (Maybe Integer)

The number of recovery points that are stored in a backup vault.

backupVaultListMember_backupVaultArn :: Lens' BackupVaultListMember (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

backupVaultListMember_encryptionKeyArn :: Lens' BackupVaultListMember (Maybe Text)

The server-side encryption key that is used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

backupVaultListMember_creationDate :: Lens' BackupVaultListMember (Maybe UTCTime)

The date and time a resource backup is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupVaultListMember_backupVaultName :: Lens' BackupVaultListMember (Maybe Text)

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

backupVaultListMember_minRetentionDays :: Lens' BackupVaultListMember (Maybe Integer)

The Backup Vault Lock setting that specifies the minimum retention period that the vault retains its recovery points. If this parameter is not specified, Vault Lock does not enforce a minimum retention period.

If specified, any backup or copy job to the vault must have a lifecycle policy with a retention period equal to or longer than the minimum retention period. If the job's retention period is shorter than that minimum retention period, then the vault fails the backup or copy job, and you should either modify your lifecycle settings or use a different vault. Recovery points already stored in the vault prior to Vault Lock are not affected.