Network.AWS.Backup.Types.Framework

Description

Documentation

data Framework

Contains detailed information about a framework. Frameworks contain controls, which evaluate and report on your backup events and resources. Frameworks generate daily compliance results.

See: newFramework smart constructor.

Constructors

Framework'

Fields

creationTime :: Maybe POSIX

The date and time that a framework is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

frameworkDescription :: Maybe Text

An optional description of the framework with a maximum 1,024 characters.

numberOfControls :: Maybe Int

The number of controls contained by the framework.

frameworkArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

deploymentStatus :: Maybe Text

The deployment status of a framework. The statuses are:

CREATE_IN_PROGRESS | UPDATE_IN_PROGRESS | DELETE_IN_PROGRESS | COMPLETED | FAILED

frameworkName :: Maybe Text

The unique name of a framework. This name is between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

newFramework :: Framework

Create a value of Framework with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Framework', framework_creationTime - The date and time that a framework is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:frameworkDescription:Framework', framework_frameworkDescription - An optional description of the framework with a maximum 1,024 characters.

$sel:numberOfControls:Framework', framework_numberOfControls - The number of controls contained by the framework.

$sel:frameworkArn:Framework', framework_frameworkArn - An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

$sel:deploymentStatus:Framework', framework_deploymentStatus - The deployment status of a framework. The statuses are:

CREATE_IN_PROGRESS | UPDATE_IN_PROGRESS | DELETE_IN_PROGRESS | COMPLETED | FAILED

$sel:frameworkName:Framework', framework_frameworkName - The unique name of a framework. This name is between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

framework_creationTime :: Lens' Framework (Maybe UTCTime)

The date and time that a framework is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

framework_frameworkDescription :: Lens' Framework (Maybe Text)

An optional description of the framework with a maximum 1,024 characters.

framework_numberOfControls :: Lens' Framework (Maybe Int)

The number of controls contained by the framework.

framework_frameworkArn :: Lens' Framework (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

framework_deploymentStatus :: Lens' Framework (Maybe Text)

The deployment status of a framework. The statuses are:

CREATE_IN_PROGRESS | UPDATE_IN_PROGRESS | DELETE_IN_PROGRESS | COMPLETED | FAILED

framework_frameworkName :: Lens' Framework (Maybe Text)

The unique name of a framework. This name is between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).