Network.AWS.Backup.Types.RecoveryPointByBackupVault

Description

Documentation

data RecoveryPointByBackupVault

Contains detailed information about the recovery points stored in a backup vault.

See: newRecoveryPointByBackupVault smart constructor.

Constructors

RecoveryPointByBackupVault'

Fields

isEncrypted :: Maybe Bool

A Boolean value that is returned as TRUE if the specified recovery point is encrypted, or FALSE if the recovery point is not encrypted.

status :: Maybe RecoveryPointStatus

A status code specifying the state of the recovery point.

iamRoleArn :: Maybe Text

Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

resourceType :: Maybe Text

The type of Amazon Web Services resource saved as a recovery point; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database. For Windows Volume Shadow Copy Service (VSS) backups, the only supported resource type is Amazon EC2.

createdBy :: Maybe RecoveryPointCreator

Contains identifying information about the creation of a recovery point, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan that is used to create it.

calculatedLifecycle :: Maybe CalculatedLifecycle

A CalculatedLifecycle object containing DeleteAt and MoveToColdStorageAt timestamps.

lifecycle :: Maybe Lifecycle

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

backupVaultArn :: Maybe Text

An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

sourceBackupVaultArn :: Maybe Text

The backup vault where the recovery point was originally copied from. If the recovery point is restored to the same account this value will be null.

lastRestoreTime :: Maybe POSIX

The date and time a recovery point was last restored, in Unix format and Coordinated Universal Time (UTC). The value of LastRestoreTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

resourceArn :: Maybe Text

An ARN that uniquely identifies a resource. The format of the ARN depends on the resource type.

statusMessage :: Maybe Text

A message explaining the reason of the recovery point deletion failure.

recoveryPointArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

encryptionKeyArn :: Maybe Text

The server-side encryption key that is used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

backupSizeInBytes :: Maybe Integer

The size, in bytes, of a backup.

creationDate :: Maybe POSIX

The date and time a recovery point is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

completionDate :: Maybe POSIX

The date and time a job to restore a recovery point is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backupVaultName :: Maybe Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

newRecoveryPointByBackupVault :: RecoveryPointByBackupVault

Create a value of RecoveryPointByBackupVault with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isEncrypted:RecoveryPointByBackupVault', recoveryPointByBackupVault_isEncrypted - A Boolean value that is returned as TRUE if the specified recovery point is encrypted, or FALSE if the recovery point is not encrypted.

$sel:status:RecoveryPointByBackupVault', recoveryPointByBackupVault_status - A status code specifying the state of the recovery point.

$sel:iamRoleArn:RecoveryPointByBackupVault', recoveryPointByBackupVault_iamRoleArn - Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

$sel:resourceType:RecoveryPointByBackupVault', recoveryPointByBackupVault_resourceType - The type of Amazon Web Services resource saved as a recovery point; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database. For Windows Volume Shadow Copy Service (VSS) backups, the only supported resource type is Amazon EC2.

$sel:createdBy:RecoveryPointByBackupVault', recoveryPointByBackupVault_createdBy - Contains identifying information about the creation of a recovery point, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan that is used to create it.

$sel:calculatedLifecycle:RecoveryPointByBackupVault', recoveryPointByBackupVault_calculatedLifecycle - A CalculatedLifecycle object containing DeleteAt and MoveToColdStorageAt timestamps.

$sel:lifecycle:RecoveryPointByBackupVault', recoveryPointByBackupVault_lifecycle - The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

$sel:backupVaultArn:RecoveryPointByBackupVault', recoveryPointByBackupVault_backupVaultArn - An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:sourceBackupVaultArn:RecoveryPointByBackupVault', recoveryPointByBackupVault_sourceBackupVaultArn - The backup vault where the recovery point was originally copied from. If the recovery point is restored to the same account this value will be null.

$sel:lastRestoreTime:RecoveryPointByBackupVault', recoveryPointByBackupVault_lastRestoreTime - The date and time a recovery point was last restored, in Unix format and Coordinated Universal Time (UTC). The value of LastRestoreTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:resourceArn:RecoveryPointByBackupVault', recoveryPointByBackupVault_resourceArn - An ARN that uniquely identifies a resource. The format of the ARN depends on the resource type.

$sel:statusMessage:RecoveryPointByBackupVault', recoveryPointByBackupVault_statusMessage - A message explaining the reason of the recovery point deletion failure.

$sel:recoveryPointArn:RecoveryPointByBackupVault', recoveryPointByBackupVault_recoveryPointArn - An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:encryptionKeyArn:RecoveryPointByBackupVault', recoveryPointByBackupVault_encryptionKeyArn - The server-side encryption key that is used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

$sel:backupSizeInBytes:RecoveryPointByBackupVault', recoveryPointByBackupVault_backupSizeInBytes - The size, in bytes, of a backup.

$sel:creationDate:RecoveryPointByBackupVault', recoveryPointByBackupVault_creationDate - The date and time a recovery point is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:completionDate:RecoveryPointByBackupVault', recoveryPointByBackupVault_completionDate - The date and time a job to restore a recovery point is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:backupVaultName:RecoveryPointByBackupVault', recoveryPointByBackupVault_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

recoveryPointByBackupVault_isEncrypted :: Lens' RecoveryPointByBackupVault (Maybe Bool)

A Boolean value that is returned as TRUE if the specified recovery point is encrypted, or FALSE if the recovery point is not encrypted.

recoveryPointByBackupVault_status :: Lens' RecoveryPointByBackupVault (Maybe RecoveryPointStatus)

A status code specifying the state of the recovery point.

recoveryPointByBackupVault_iamRoleArn :: Lens' RecoveryPointByBackupVault (Maybe Text)

Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

recoveryPointByBackupVault_resourceType :: Lens' RecoveryPointByBackupVault (Maybe Text)

The type of Amazon Web Services resource saved as a recovery point; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database. For Windows Volume Shadow Copy Service (VSS) backups, the only supported resource type is Amazon EC2.

recoveryPointByBackupVault_createdBy :: Lens' RecoveryPointByBackupVault (Maybe RecoveryPointCreator)

Contains identifying information about the creation of a recovery point, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan that is used to create it.

recoveryPointByBackupVault_calculatedLifecycle :: Lens' RecoveryPointByBackupVault (Maybe CalculatedLifecycle)

A CalculatedLifecycle object containing DeleteAt and MoveToColdStorageAt timestamps.

recoveryPointByBackupVault_lifecycle :: Lens' RecoveryPointByBackupVault (Maybe Lifecycle)

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

recoveryPointByBackupVault_backupVaultArn :: Lens' RecoveryPointByBackupVault (Maybe Text)

An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

recoveryPointByBackupVault_sourceBackupVaultArn :: Lens' RecoveryPointByBackupVault (Maybe Text)

The backup vault where the recovery point was originally copied from. If the recovery point is restored to the same account this value will be null.

recoveryPointByBackupVault_lastRestoreTime :: Lens' RecoveryPointByBackupVault (Maybe UTCTime)

The date and time a recovery point was last restored, in Unix format and Coordinated Universal Time (UTC). The value of LastRestoreTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

recoveryPointByBackupVault_resourceArn :: Lens' RecoveryPointByBackupVault (Maybe Text)

An ARN that uniquely identifies a resource. The format of the ARN depends on the resource type.

recoveryPointByBackupVault_statusMessage :: Lens' RecoveryPointByBackupVault (Maybe Text)

A message explaining the reason of the recovery point deletion failure.

recoveryPointByBackupVault_recoveryPointArn :: Lens' RecoveryPointByBackupVault (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

recoveryPointByBackupVault_encryptionKeyArn :: Lens' RecoveryPointByBackupVault (Maybe Text)

The server-side encryption key that is used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

recoveryPointByBackupVault_backupSizeInBytes :: Lens' RecoveryPointByBackupVault (Maybe Integer)

The size, in bytes, of a backup.

recoveryPointByBackupVault_creationDate :: Lens' RecoveryPointByBackupVault (Maybe UTCTime)

The date and time a recovery point is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

recoveryPointByBackupVault_completionDate :: Lens' RecoveryPointByBackupVault (Maybe UTCTime)

The date and time a job to restore a recovery point is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

recoveryPointByBackupVault_backupVaultName :: Lens' RecoveryPointByBackupVault (Maybe Text)

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.