Network.AWS.Backup.Types.RecoveryPointCreator

Description

Documentation

data RecoveryPointCreator

Contains information about the backup plan and rule that Backup used to initiate the recovery point backup.

See: newRecoveryPointCreator smart constructor.

Constructors

RecoveryPointCreator'

Fields

backupPlanId :: Maybe Text

Uniquely identifies a backup plan.

backupPlanArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

backupPlanVersion :: Maybe Text

Version IDs are unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. They cannot be edited.

backupRuleId :: Maybe Text

Uniquely identifies a rule used to schedule the backup of a selection of resources.

newRecoveryPointCreator :: RecoveryPointCreator

Create a value of RecoveryPointCreator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupPlanId:RecoveryPointCreator', recoveryPointCreator_backupPlanId - Uniquely identifies a backup plan.

$sel:backupPlanArn:RecoveryPointCreator', recoveryPointCreator_backupPlanArn - An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

$sel:backupPlanVersion:RecoveryPointCreator', recoveryPointCreator_backupPlanVersion - Version IDs are unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. They cannot be edited.

$sel:backupRuleId:RecoveryPointCreator', recoveryPointCreator_backupRuleId - Uniquely identifies a rule used to schedule the backup of a selection of resources.

recoveryPointCreator_backupPlanId :: Lens' RecoveryPointCreator (Maybe Text)

Uniquely identifies a backup plan.

recoveryPointCreator_backupPlanArn :: Lens' RecoveryPointCreator (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a backup plan; for example, arn:aws:backup:us-east-1:123456789012:plan:8F81F553-3A74-4A3F-B93D-B3360DC80C50.

recoveryPointCreator_backupPlanVersion :: Lens' RecoveryPointCreator (Maybe Text)

Version IDs are unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. They cannot be edited.

recoveryPointCreator_backupRuleId :: Lens' RecoveryPointCreator (Maybe Text)

Uniquely identifies a rule used to schedule the backup of a selection of resources.