Network.AWS.Backup.Types.ReportDestination

Description

Documentation

data ReportDestination

Contains information from your report job about your report destination.

See: newReportDestination smart constructor.

Constructors

ReportDestination'

Fields

s3Keys :: Maybe [Text]

The object key that uniquely identifies your reports in your S3 bucket.

s3BucketName :: Maybe Text

The unique name of the Amazon S3 bucket that receives your reports.

newReportDestination :: ReportDestination

Create a value of ReportDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Keys:ReportDestination', reportDestination_s3Keys - The object key that uniquely identifies your reports in your S3 bucket.

$sel:s3BucketName:ReportDestination', reportDestination_s3BucketName - The unique name of the Amazon S3 bucket that receives your reports.

reportDestination_s3Keys :: Lens' ReportDestination (Maybe [Text])

The object key that uniquely identifies your reports in your S3 bucket.

reportDestination_s3BucketName :: Lens' ReportDestination (Maybe Text)

The unique name of the Amazon S3 bucket that receives your reports.