Network.AWS.Backup.Types.ReportPlan

Description

Documentation

data ReportPlan

Contains detailed information about a report plan.

See: newReportPlan smart constructor.

Constructors

ReportPlan'

Fields

creationTime :: Maybe POSIX

The date and time that a report plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

reportPlanName :: Maybe Text

The unique name of the report plan. This name is between 1 and 256 characters starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

reportPlanArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

reportSetting :: Maybe ReportSetting

Identifies the report template for the report. Reports are built using a report template. The report templates are:

RESOURCE_COMPLIANCE_REPORT | CONTROL_COMPLIANCE_REPORT | BACKUP_JOB_REPORT | COPY_JOB_REPORT | RESTORE_JOB_REPORT

If the report template is RESOURCE_COMPLIANCE_REPORT or CONTROL_COMPLIANCE_REPORT, this API resource also describes the report coverage by Amazon Web Services Regions and frameworks.

reportPlanDescription :: Maybe Text

An optional description of the report plan with a maximum 1,024 characters.

lastAttemptedExecutionTime :: Maybe POSIX

The date and time that a report job associated with this report plan last attempted to run, in Unix format and Coordinated Universal Time (UTC). The value of LastAttemptedExecutionTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

deploymentStatus :: Maybe Text

The deployment status of a report plan. The statuses are:

CREATE_IN_PROGRESS | UPDATE_IN_PROGRESS | DELETE_IN_PROGRESS | COMPLETED

lastSuccessfulExecutionTime :: Maybe POSIX

The date and time that a report job associated with this report plan last successfully ran, in Unix format and Coordinated Universal Time (UTC). The value of LastSuccessfulExecutionTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

reportDeliveryChannel :: Maybe ReportDeliveryChannel

Contains information about where and how to deliver your reports, specifically your Amazon S3 bucket name, S3 key prefix, and the formats of your reports.

newReportPlan :: ReportPlan

Create a value of ReportPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ReportPlan', reportPlan_creationTime - The date and time that a report plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:reportPlanName:ReportPlan', reportPlan_reportPlanName - The unique name of the report plan. This name is between 1 and 256 characters starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

$sel:reportPlanArn:ReportPlan', reportPlan_reportPlanArn - An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

$sel:reportSetting:ReportPlan', reportPlan_reportSetting - Identifies the report template for the report. Reports are built using a report template. The report templates are:

RESOURCE_COMPLIANCE_REPORT | CONTROL_COMPLIANCE_REPORT | BACKUP_JOB_REPORT | COPY_JOB_REPORT | RESTORE_JOB_REPORT

If the report template is RESOURCE_COMPLIANCE_REPORT or CONTROL_COMPLIANCE_REPORT, this API resource also describes the report coverage by Amazon Web Services Regions and frameworks.

$sel:reportPlanDescription:ReportPlan', reportPlan_reportPlanDescription - An optional description of the report plan with a maximum 1,024 characters.

$sel:lastAttemptedExecutionTime:ReportPlan', reportPlan_lastAttemptedExecutionTime - The date and time that a report job associated with this report plan last attempted to run, in Unix format and Coordinated Universal Time (UTC). The value of LastAttemptedExecutionTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:deploymentStatus:ReportPlan', reportPlan_deploymentStatus - The deployment status of a report plan. The statuses are:

CREATE_IN_PROGRESS | UPDATE_IN_PROGRESS | DELETE_IN_PROGRESS | COMPLETED

$sel:lastSuccessfulExecutionTime:ReportPlan', reportPlan_lastSuccessfulExecutionTime - The date and time that a report job associated with this report plan last successfully ran, in Unix format and Coordinated Universal Time (UTC). The value of LastSuccessfulExecutionTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:reportDeliveryChannel:ReportPlan', reportPlan_reportDeliveryChannel - Contains information about where and how to deliver your reports, specifically your Amazon S3 bucket name, S3 key prefix, and the formats of your reports.

reportPlan_creationTime :: Lens' ReportPlan (Maybe UTCTime)

The date and time that a report plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

reportPlan_reportPlanName :: Lens' ReportPlan (Maybe Text)

The unique name of the report plan. This name is between 1 and 256 characters starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

reportPlan_reportPlanArn :: Lens' ReportPlan (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

reportPlan_reportSetting :: Lens' ReportPlan (Maybe ReportSetting)

Identifies the report template for the report. Reports are built using a report template. The report templates are:

RESOURCE_COMPLIANCE_REPORT | CONTROL_COMPLIANCE_REPORT | BACKUP_JOB_REPORT | COPY_JOB_REPORT | RESTORE_JOB_REPORT

If the report template is RESOURCE_COMPLIANCE_REPORT or CONTROL_COMPLIANCE_REPORT, this API resource also describes the report coverage by Amazon Web Services Regions and frameworks.

reportPlan_reportPlanDescription :: Lens' ReportPlan (Maybe Text)

An optional description of the report plan with a maximum 1,024 characters.

reportPlan_lastAttemptedExecutionTime :: Lens' ReportPlan (Maybe UTCTime)

The date and time that a report job associated with this report plan last attempted to run, in Unix format and Coordinated Universal Time (UTC). The value of LastAttemptedExecutionTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

reportPlan_deploymentStatus :: Lens' ReportPlan (Maybe Text)

The deployment status of a report plan. The statuses are:

CREATE_IN_PROGRESS | UPDATE_IN_PROGRESS | DELETE_IN_PROGRESS | COMPLETED

reportPlan_lastSuccessfulExecutionTime :: Lens' ReportPlan (Maybe UTCTime)

The date and time that a report job associated with this report plan last successfully ran, in Unix format and Coordinated Universal Time (UTC). The value of LastSuccessfulExecutionTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

reportPlan_reportDeliveryChannel :: Lens' ReportPlan (Maybe ReportDeliveryChannel)

Contains information about where and how to deliver your reports, specifically your Amazon S3 bucket name, S3 key prefix, and the formats of your reports.