Network.AWS.Backup.UpdateRecoveryPointLifecycle

Description

Sets the transition lifecycle of a recovery point.

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

Does not support continuous backups.

Creating a Request

data UpdateRecoveryPointLifecycle

See: newUpdateRecoveryPointLifecycle smart constructor.

Constructors

UpdateRecoveryPointLifecycle'

Fields

lifecycle :: Maybe Lifecycle

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

backupVaultName :: Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

recoveryPointArn :: Text

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

newUpdateRecoveryPointLifecycle

Arguments

:: Text
$sel:backupVaultName:UpdateRecoveryPointLifecycle'

-> Text
$sel:recoveryPointArn:UpdateRecoveryPointLifecycle'

-> UpdateRecoveryPointLifecycle

Create a value of UpdateRecoveryPointLifecycle with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lifecycle:UpdateRecoveryPointLifecycle', updateRecoveryPointLifecycle_lifecycle - The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

$sel:backupVaultName:UpdateRecoveryPointLifecycle', updateRecoveryPointLifecycle_backupVaultName - The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

$sel:recoveryPointArn:UpdateRecoveryPointLifecycle', updateRecoveryPointLifecycle_recoveryPointArn - An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

Request Lenses

updateRecoveryPointLifecycle_lifecycle :: Lens' UpdateRecoveryPointLifecycle (Maybe Lifecycle)

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

updateRecoveryPointLifecycle_backupVaultName :: Lens' UpdateRecoveryPointLifecycle Text

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

updateRecoveryPointLifecycle_recoveryPointArn :: Lens' UpdateRecoveryPointLifecycle Text

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

Destructuring the Response

data UpdateRecoveryPointLifecycleResponse

See: newUpdateRecoveryPointLifecycleResponse smart constructor.

Constructors

UpdateRecoveryPointLifecycleResponse'

Fields

calculatedLifecycle :: Maybe CalculatedLifecycle

A CalculatedLifecycle object containing DeleteAt and MoveToColdStorageAt timestamps.

lifecycle :: Maybe Lifecycle

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

backupVaultArn :: Maybe Text

An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

recoveryPointArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

httpStatus :: Int

The response's http status code.

newUpdateRecoveryPointLifecycleResponse

Arguments

:: Int
$sel:httpStatus:UpdateRecoveryPointLifecycleResponse'

-> UpdateRecoveryPointLifecycleResponse

Create a value of UpdateRecoveryPointLifecycleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:calculatedLifecycle:UpdateRecoveryPointLifecycleResponse', updateRecoveryPointLifecycleResponse_calculatedLifecycle - A CalculatedLifecycle object containing DeleteAt and MoveToColdStorageAt timestamps.

$sel:lifecycle:UpdateRecoveryPointLifecycle', updateRecoveryPointLifecycleResponse_lifecycle - The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

$sel:backupVaultArn:UpdateRecoveryPointLifecycleResponse', updateRecoveryPointLifecycleResponse_backupVaultArn - An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

$sel:recoveryPointArn:UpdateRecoveryPointLifecycle', updateRecoveryPointLifecycleResponse_recoveryPointArn - An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

$sel:httpStatus:UpdateRecoveryPointLifecycleResponse', updateRecoveryPointLifecycleResponse_httpStatus - The response's http status code.

Response Lenses

updateRecoveryPointLifecycleResponse_calculatedLifecycle :: Lens' UpdateRecoveryPointLifecycleResponse (Maybe CalculatedLifecycle)

A CalculatedLifecycle object containing DeleteAt and MoveToColdStorageAt timestamps.

updateRecoveryPointLifecycleResponse_lifecycle :: Lens' UpdateRecoveryPointLifecycleResponse (Maybe Lifecycle)

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

updateRecoveryPointLifecycleResponse_backupVaultArn :: Lens' UpdateRecoveryPointLifecycleResponse (Maybe Text)

An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

updateRecoveryPointLifecycleResponse_recoveryPointArn :: Lens' UpdateRecoveryPointLifecycleResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

updateRecoveryPointLifecycleResponse_httpStatus :: Lens' UpdateRecoveryPointLifecycleResponse Int

The response's http status code.