Network.AWS.Backup.UpdateReportPlan

Description

Updates an existing report plan identified by its ReportPlanName with the input document in JSON format.

Creating a Request

data UpdateReportPlan

See: newUpdateReportPlan smart constructor.

Constructors

UpdateReportPlan'

Fields

idempotencyToken :: Maybe Text

A customer-chosen string that you can use to distinguish between otherwise identical calls to UpdateReportPlanInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

reportSetting :: Maybe ReportSetting

Identifies the report template for the report. Reports are built using a report template. The report templates are:

RESOURCE_COMPLIANCE_REPORT | CONTROL_COMPLIANCE_REPORT | BACKUP_JOB_REPORT | COPY_JOB_REPORT | RESTORE_JOB_REPORT

If the report template is RESOURCE_COMPLIANCE_REPORT or CONTROL_COMPLIANCE_REPORT, this API resource also describes the report coverage by Amazon Web Services Regions and frameworks.

reportPlanDescription :: Maybe Text

An optional description of the report plan with a maximum 1,024 characters.

reportDeliveryChannel :: Maybe ReportDeliveryChannel

A structure that contains information about where to deliver your reports, specifically your Amazon S3 bucket name, S3 key prefix, and the formats of your reports.

reportPlanName :: Text

The unique name of the report plan. This name is between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

newUpdateReportPlan

Arguments

:: Text
$sel:reportPlanName:UpdateReportPlan'

-> UpdateReportPlan

Create a value of UpdateReportPlan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idempotencyToken:UpdateReportPlan', updateReportPlan_idempotencyToken - A customer-chosen string that you can use to distinguish between otherwise identical calls to UpdateReportPlanInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

$sel:reportSetting:UpdateReportPlan', updateReportPlan_reportSetting - Identifies the report template for the report. Reports are built using a report template. The report templates are:

RESOURCE_COMPLIANCE_REPORT | CONTROL_COMPLIANCE_REPORT | BACKUP_JOB_REPORT | COPY_JOB_REPORT | RESTORE_JOB_REPORT

If the report template is RESOURCE_COMPLIANCE_REPORT or CONTROL_COMPLIANCE_REPORT, this API resource also describes the report coverage by Amazon Web Services Regions and frameworks.

$sel:reportPlanDescription:UpdateReportPlan', updateReportPlan_reportPlanDescription - An optional description of the report plan with a maximum 1,024 characters.

$sel:reportDeliveryChannel:UpdateReportPlan', updateReportPlan_reportDeliveryChannel - A structure that contains information about where to deliver your reports, specifically your Amazon S3 bucket name, S3 key prefix, and the formats of your reports.

$sel:reportPlanName:UpdateReportPlan', updateReportPlan_reportPlanName - The unique name of the report plan. This name is between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

Request Lenses

updateReportPlan_idempotencyToken :: Lens' UpdateReportPlan (Maybe Text)

A customer-chosen string that you can use to distinguish between otherwise identical calls to UpdateReportPlanInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

updateReportPlan_reportSetting :: Lens' UpdateReportPlan (Maybe ReportSetting)

Identifies the report template for the report. Reports are built using a report template. The report templates are:

RESOURCE_COMPLIANCE_REPORT | CONTROL_COMPLIANCE_REPORT | BACKUP_JOB_REPORT | COPY_JOB_REPORT | RESTORE_JOB_REPORT

If the report template is RESOURCE_COMPLIANCE_REPORT or CONTROL_COMPLIANCE_REPORT, this API resource also describes the report coverage by Amazon Web Services Regions and frameworks.

updateReportPlan_reportPlanDescription :: Lens' UpdateReportPlan (Maybe Text)

An optional description of the report plan with a maximum 1,024 characters.

updateReportPlan_reportDeliveryChannel :: Lens' UpdateReportPlan (Maybe ReportDeliveryChannel)

A structure that contains information about where to deliver your reports, specifically your Amazon S3 bucket name, S3 key prefix, and the formats of your reports.

updateReportPlan_reportPlanName :: Lens' UpdateReportPlan Text

The unique name of the report plan. This name is between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

Destructuring the Response

data UpdateReportPlanResponse

See: newUpdateReportPlanResponse smart constructor.

Constructors

UpdateReportPlanResponse'

Fields

creationTime :: Maybe POSIX

The date and time that a report plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

reportPlanName :: Maybe Text

The unique name of the report plan.

reportPlanArn :: Maybe Text

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

httpStatus :: Int

The response's http status code.

newUpdateReportPlanResponse

Arguments

:: Int
$sel:httpStatus:UpdateReportPlanResponse'

-> UpdateReportPlanResponse

Create a value of UpdateReportPlanResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:UpdateReportPlanResponse', updateReportPlanResponse_creationTime - The date and time that a report plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

$sel:reportPlanName:UpdateReportPlan', updateReportPlanResponse_reportPlanName - The unique name of the report plan.

$sel:reportPlanArn:UpdateReportPlanResponse', updateReportPlanResponse_reportPlanArn - An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

$sel:httpStatus:UpdateReportPlanResponse', updateReportPlanResponse_httpStatus - The response's http status code.

Response Lenses

updateReportPlanResponse_creationTime :: Lens' UpdateReportPlanResponse (Maybe UTCTime)

The date and time that a report plan is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

updateReportPlanResponse_reportPlanName :: Lens' UpdateReportPlanResponse (Maybe Text)

The unique name of the report plan.

updateReportPlanResponse_reportPlanArn :: Lens' UpdateReportPlanResponse (Maybe Text)

An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

updateReportPlanResponse_httpStatus :: Lens' UpdateReportPlanResponse Int

The response's http status code.