Network.AWS.Batch.DeleteComputeEnvironment

Description

Deletes an Batch compute environment.

Before you can delete a compute environment, you must set its state to DISABLED with the UpdateComputeEnvironment API operation and disassociate it from any job queues with the UpdateJobQueue API operation. Compute environments that use Fargate resources must terminate all active jobs on that compute environment before deleting the compute environment. If this isn't done, the compute environment enters an invalid state.

Creating a Request

data DeleteComputeEnvironment

Contains the parameters for DeleteComputeEnvironment.

See: newDeleteComputeEnvironment smart constructor.

Constructors

DeleteComputeEnvironment'

Fields

computeEnvironment :: Text

The name or Amazon Resource Name (ARN) of the compute environment to delete.

newDeleteComputeEnvironment

Arguments

:: Text
$sel:computeEnvironment:DeleteComputeEnvironment'

-> DeleteComputeEnvironment

Create a value of DeleteComputeEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:computeEnvironment:DeleteComputeEnvironment', deleteComputeEnvironment_computeEnvironment - The name or Amazon Resource Name (ARN) of the compute environment to delete.

Request Lenses

deleteComputeEnvironment_computeEnvironment :: Lens' DeleteComputeEnvironment Text

The name or Amazon Resource Name (ARN) of the compute environment to delete.

Destructuring the Response

data DeleteComputeEnvironmentResponse

See: newDeleteComputeEnvironmentResponse smart constructor.

Constructors

DeleteComputeEnvironmentResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteComputeEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:DeleteComputeEnvironmentResponse'

-> DeleteComputeEnvironmentResponse

Create a value of DeleteComputeEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteComputeEnvironmentResponse', deleteComputeEnvironmentResponse_httpStatus - The response's http status code.

Response Lenses

deleteComputeEnvironmentResponse_httpStatus :: Lens' DeleteComputeEnvironmentResponse Int

The response's http status code.