Network.AWS.Batch.DescribeJobDefinitions

Description

Describes a list of job definitions. You can specify a status (such as ACTIVE) to only return job definitions that match that status.

This operation returns paginated results.

Creating a Request

data DescribeJobDefinitions

Contains the parameters for DescribeJobDefinitions.

See: newDescribeJobDefinitions smart constructor.

Constructors

DescribeJobDefinitions'

Fields

nextToken :: Maybe Text

The nextToken value returned from a previous paginated DescribeJobDefinitions request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

This token should be treated as an opaque identifier that's only used to retrieve the next items in a list and not for other programmatic purposes.

status :: Maybe Text

The status used to filter job definitions.

jobDefinitions :: Maybe [Text]

A list of up to 100 job definitions. Each entry in the list can either be an ARN of the form arn:aws:batch:${Region}:${Account}:job-definition/${JobDefinitionName}:${Revision} or a short version using the form ${JobDefinitionName}:${Revision}.

maxResults :: Maybe Int

The maximum number of results returned by DescribeJobDefinitions in paginated output. When this parameter is used, DescribeJobDefinitions only returns maxResults results in a single page and a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeJobDefinitions request with the returned nextToken value. This value can be between 1 and 100. If this parameter isn't used, then DescribeJobDefinitions returns up to 100 results and a nextToken value if applicable.

jobDefinitionName :: Maybe Text

The name of the job definition to describe.

newDescribeJobDefinitions :: DescribeJobDefinitions

Create a value of DescribeJobDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeJobDefinitions', describeJobDefinitions_nextToken - The nextToken value returned from a previous paginated DescribeJobDefinitions request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

This token should be treated as an opaque identifier that's only used to retrieve the next items in a list and not for other programmatic purposes.

$sel:status:DescribeJobDefinitions', describeJobDefinitions_status - The status used to filter job definitions.

$sel:jobDefinitions:DescribeJobDefinitions', describeJobDefinitions_jobDefinitions - A list of up to 100 job definitions. Each entry in the list can either be an ARN of the form arn:aws:batch:${Region}:${Account}:job-definition/${JobDefinitionName}:${Revision} or a short version using the form ${JobDefinitionName}:${Revision}.

$sel:maxResults:DescribeJobDefinitions', describeJobDefinitions_maxResults - The maximum number of results returned by DescribeJobDefinitions in paginated output. When this parameter is used, DescribeJobDefinitions only returns maxResults results in a single page and a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeJobDefinitions request with the returned nextToken value. This value can be between 1 and 100. If this parameter isn't used, then DescribeJobDefinitions returns up to 100 results and a nextToken value if applicable.

$sel:jobDefinitionName:DescribeJobDefinitions', describeJobDefinitions_jobDefinitionName - The name of the job definition to describe.

Request Lenses

describeJobDefinitions_nextToken :: Lens' DescribeJobDefinitions (Maybe Text)

The nextToken value returned from a previous paginated DescribeJobDefinitions request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

This token should be treated as an opaque identifier that's only used to retrieve the next items in a list and not for other programmatic purposes.

describeJobDefinitions_status :: Lens' DescribeJobDefinitions (Maybe Text)

The status used to filter job definitions.

describeJobDefinitions_jobDefinitions :: Lens' DescribeJobDefinitions (Maybe [Text])

A list of up to 100 job definitions. Each entry in the list can either be an ARN of the form arn:aws:batch:${Region}:${Account}:job-definition/${JobDefinitionName}:${Revision} or a short version using the form ${JobDefinitionName}:${Revision}.

describeJobDefinitions_maxResults :: Lens' DescribeJobDefinitions (Maybe Int)

The maximum number of results returned by DescribeJobDefinitions in paginated output. When this parameter is used, DescribeJobDefinitions only returns maxResults results in a single page and a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeJobDefinitions request with the returned nextToken value. This value can be between 1 and 100. If this parameter isn't used, then DescribeJobDefinitions returns up to 100 results and a nextToken value if applicable.

describeJobDefinitions_jobDefinitionName :: Lens' DescribeJobDefinitions (Maybe Text)

The name of the job definition to describe.

Destructuring the Response

data DescribeJobDefinitionsResponse

See: newDescribeJobDefinitionsResponse smart constructor.

Constructors

DescribeJobDefinitionsResponse'

Fields

nextToken :: Maybe Text

The nextToken value to include in a future DescribeJobDefinitions request. When the results of a DescribeJobDefinitions request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

jobDefinitions :: Maybe [JobDefinition]

The list of job definitions.

httpStatus :: Int

The response's http status code.

newDescribeJobDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeJobDefinitionsResponse'

-> DescribeJobDefinitionsResponse

Create a value of DescribeJobDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeJobDefinitions', describeJobDefinitionsResponse_nextToken - The nextToken value to include in a future DescribeJobDefinitions request. When the results of a DescribeJobDefinitions request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

$sel:jobDefinitions:DescribeJobDefinitions', describeJobDefinitionsResponse_jobDefinitions - The list of job definitions.

$sel:httpStatus:DescribeJobDefinitionsResponse', describeJobDefinitionsResponse_httpStatus - The response's http status code.

Response Lenses

describeJobDefinitionsResponse_nextToken :: Lens' DescribeJobDefinitionsResponse (Maybe Text)

The nextToken value to include in a future DescribeJobDefinitions request. When the results of a DescribeJobDefinitions request exceed maxResults, this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

describeJobDefinitionsResponse_jobDefinitions :: Lens' DescribeJobDefinitionsResponse (Maybe [JobDefinition])

The list of job definitions.

describeJobDefinitionsResponse_httpStatus :: Lens' DescribeJobDefinitionsResponse Int

The response's http status code.