Network.AWS.Batch.RegisterJobDefinition

Description

Registers an Batch job definition.

Creating a Request

data RegisterJobDefinition

Contains the parameters for RegisterJobDefinition.

See: newRegisterJobDefinition smart constructor.

Constructors

RegisterJobDefinition'

Fields

platformCapabilities :: Maybe [PlatformCapability]

The platform capabilities required by the job definition. If no value is specified, it defaults to EC2. To run the job on Fargate resources, specify FARGATE.

timeout :: Maybe JobTimeout

The timeout configuration for jobs that are submitted with this job definition, after which Batch terminates your jobs if they have not finished. If a job is terminated due to a timeout, it isn't retried. The minimum value for the timeout is 60 seconds. Any timeout configuration that's specified during a SubmitJob operation overrides the timeout configuration defined here. For more information, see Job Timeouts in the Batch User Guide.

nodeProperties :: Maybe NodeProperties

An object with various properties specific to multi-node parallel jobs. If you specify node properties for a job, it becomes a multi-node parallel job. For more information, see Multi-node Parallel Jobs in the Batch User Guide. If the job definition's type parameter is container, then you must specify either containerProperties or nodeProperties.

If the job runs on Fargate resources, then you must not specify nodeProperties; use containerProperties instead.

tags :: Maybe (HashMap Text Text)

The tags that you apply to the job definition to help you categorize and organize your resources. Each tag consists of a key and an optional value. For more information, see Tagging Amazon Web Services Resources in Batch User Guide.

containerProperties :: Maybe ContainerProperties

An object with various properties specific to single-node container-based jobs. If the job definition's type parameter is container, then you must specify either containerProperties or nodeProperties.

If the job runs on Fargate resources, then you must not specify nodeProperties; use only containerProperties.

retryStrategy :: Maybe RetryStrategy

The retry strategy to use for failed jobs that are submitted with this job definition. Any retry strategy that's specified during a SubmitJob operation overrides the retry strategy defined here. If a job is terminated due to a timeout, it isn't retried.

parameters :: Maybe (HashMap Text Text)

Default parameter substitution placeholders to set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

propagateTags :: Maybe Bool

Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags are not propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state.

jobDefinitionName :: Text

The name of the job definition to register. Up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

type' :: JobDefinitionType

The type of job definition. For more information about multi-node parallel jobs, see Creating a multi-node parallel job definition in the Batch User Guide.

If the job is run on Fargate resources, then multinode isn't supported.

newRegisterJobDefinition

Arguments

:: Text
$sel:jobDefinitionName:RegisterJobDefinition'

-> JobDefinitionType
$sel:type':RegisterJobDefinition'

-> RegisterJobDefinition

Create a value of RegisterJobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platformCapabilities:RegisterJobDefinition', registerJobDefinition_platformCapabilities - The platform capabilities required by the job definition. If no value is specified, it defaults to EC2. To run the job on Fargate resources, specify FARGATE.

$sel:timeout:RegisterJobDefinition', registerJobDefinition_timeout - The timeout configuration for jobs that are submitted with this job definition, after which Batch terminates your jobs if they have not finished. If a job is terminated due to a timeout, it isn't retried. The minimum value for the timeout is 60 seconds. Any timeout configuration that's specified during a SubmitJob operation overrides the timeout configuration defined here. For more information, see Job Timeouts in the Batch User Guide.

$sel:nodeProperties:RegisterJobDefinition', registerJobDefinition_nodeProperties - An object with various properties specific to multi-node parallel jobs. If you specify node properties for a job, it becomes a multi-node parallel job. For more information, see Multi-node Parallel Jobs in the Batch User Guide. If the job definition's type parameter is container, then you must specify either containerProperties or nodeProperties.

If the job runs on Fargate resources, then you must not specify nodeProperties; use containerProperties instead.

$sel:tags:RegisterJobDefinition', registerJobDefinition_tags - The tags that you apply to the job definition to help you categorize and organize your resources. Each tag consists of a key and an optional value. For more information, see Tagging Amazon Web Services Resources in Batch User Guide.

$sel:containerProperties:RegisterJobDefinition', registerJobDefinition_containerProperties - An object with various properties specific to single-node container-based jobs. If the job definition's type parameter is container, then you must specify either containerProperties or nodeProperties.

If the job runs on Fargate resources, then you must not specify nodeProperties; use only containerProperties.

$sel:retryStrategy:RegisterJobDefinition', registerJobDefinition_retryStrategy - The retry strategy to use for failed jobs that are submitted with this job definition. Any retry strategy that's specified during a SubmitJob operation overrides the retry strategy defined here. If a job is terminated due to a timeout, it isn't retried.

$sel:parameters:RegisterJobDefinition', registerJobDefinition_parameters - Default parameter substitution placeholders to set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

$sel:propagateTags:RegisterJobDefinition', registerJobDefinition_propagateTags - Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags are not propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state.

$sel:jobDefinitionName:RegisterJobDefinition', registerJobDefinition_jobDefinitionName - The name of the job definition to register. Up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

$sel:type':RegisterJobDefinition', registerJobDefinition_type - The type of job definition. For more information about multi-node parallel jobs, see Creating a multi-node parallel job definition in the Batch User Guide.

If the job is run on Fargate resources, then multinode isn't supported.

Request Lenses

registerJobDefinition_platformCapabilities :: Lens' RegisterJobDefinition (Maybe [PlatformCapability])

The platform capabilities required by the job definition. If no value is specified, it defaults to EC2. To run the job on Fargate resources, specify FARGATE.

registerJobDefinition_timeout :: Lens' RegisterJobDefinition (Maybe JobTimeout)

The timeout configuration for jobs that are submitted with this job definition, after which Batch terminates your jobs if they have not finished. If a job is terminated due to a timeout, it isn't retried. The minimum value for the timeout is 60 seconds. Any timeout configuration that's specified during a SubmitJob operation overrides the timeout configuration defined here. For more information, see Job Timeouts in the Batch User Guide.

registerJobDefinition_nodeProperties :: Lens' RegisterJobDefinition (Maybe NodeProperties)

An object with various properties specific to multi-node parallel jobs. If you specify node properties for a job, it becomes a multi-node parallel job. For more information, see Multi-node Parallel Jobs in the Batch User Guide. If the job definition's type parameter is container, then you must specify either containerProperties or nodeProperties.

If the job runs on Fargate resources, then you must not specify nodeProperties; use containerProperties instead.

registerJobDefinition_tags :: Lens' RegisterJobDefinition (Maybe (HashMap Text Text))

The tags that you apply to the job definition to help you categorize and organize your resources. Each tag consists of a key and an optional value. For more information, see Tagging Amazon Web Services Resources in Batch User Guide.

registerJobDefinition_containerProperties :: Lens' RegisterJobDefinition (Maybe ContainerProperties)

An object with various properties specific to single-node container-based jobs. If the job definition's type parameter is container, then you must specify either containerProperties or nodeProperties.

If the job runs on Fargate resources, then you must not specify nodeProperties; use only containerProperties.

registerJobDefinition_retryStrategy :: Lens' RegisterJobDefinition (Maybe RetryStrategy)

The retry strategy to use for failed jobs that are submitted with this job definition. Any retry strategy that's specified during a SubmitJob operation overrides the retry strategy defined here. If a job is terminated due to a timeout, it isn't retried.

registerJobDefinition_parameters :: Lens' RegisterJobDefinition (Maybe (HashMap Text Text))

Default parameter substitution placeholders to set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

registerJobDefinition_propagateTags :: Lens' RegisterJobDefinition (Maybe Bool)

Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags are not propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state.

registerJobDefinition_jobDefinitionName :: Lens' RegisterJobDefinition Text

The name of the job definition to register. Up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

registerJobDefinition_type :: Lens' RegisterJobDefinition JobDefinitionType

The type of job definition. For more information about multi-node parallel jobs, see Creating a multi-node parallel job definition in the Batch User Guide.

If the job is run on Fargate resources, then multinode isn't supported.

Destructuring the Response

data RegisterJobDefinitionResponse

See: newRegisterJobDefinitionResponse smart constructor.

Constructors

RegisterJobDefinitionResponse'

Fields

httpStatus :: Int

The response's http status code.

jobDefinitionName :: Text

The name of the job definition.

jobDefinitionArn :: Text

The Amazon Resource Name (ARN) of the job definition.

revision :: Int

The revision of the job definition.

newRegisterJobDefinitionResponse

Arguments

:: Int
$sel:httpStatus:RegisterJobDefinitionResponse'

-> Text
$sel:jobDefinitionName:RegisterJobDefinition'

-> Text
$sel:jobDefinitionArn:RegisterJobDefinitionResponse'

-> Int
$sel:revision:RegisterJobDefinitionResponse'

-> RegisterJobDefinitionResponse

Create a value of RegisterJobDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RegisterJobDefinitionResponse', registerJobDefinitionResponse_httpStatus - The response's http status code.

$sel:jobDefinitionName:RegisterJobDefinition', registerJobDefinitionResponse_jobDefinitionName - The name of the job definition.

$sel:jobDefinitionArn:RegisterJobDefinitionResponse', registerJobDefinitionResponse_jobDefinitionArn - The Amazon Resource Name (ARN) of the job definition.

$sel:revision:RegisterJobDefinitionResponse', registerJobDefinitionResponse_revision - The revision of the job definition.

Response Lenses

registerJobDefinitionResponse_httpStatus :: Lens' RegisterJobDefinitionResponse Int

The response's http status code.

registerJobDefinitionResponse_jobDefinitionName :: Lens' RegisterJobDefinitionResponse Text

The name of the job definition.

registerJobDefinitionResponse_jobDefinitionArn :: Lens' RegisterJobDefinitionResponse Text

The Amazon Resource Name (ARN) of the job definition.

registerJobDefinitionResponse_revision :: Lens' RegisterJobDefinitionResponse Int

The revision of the job definition.