Network.AWS.Batch.Types.AttemptContainerDetail

Description

Documentation

data AttemptContainerDetail

An object representing the details of a container that's part of a job attempt.

See: newAttemptContainerDetail smart constructor.

Constructors

AttemptContainerDetail'

Fields

logStreamName :: Maybe Text

The name of the CloudWatch Logs log stream associated with the container. The log group for Batch jobs is /aws/batch/job. Each container attempt receives a log stream name when they reach the RUNNING status.

containerInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon ECS container instance that hosts the job attempt.

exitCode :: Maybe Int

The exit code for the job attempt. A non-zero exit code is considered a failure.

reason :: Maybe Text

A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

networkInterfaces :: Maybe [NetworkInterface]

The network interfaces associated with the job attempt.

taskArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon ECS task that's associated with the job attempt. Each container attempt receives a task ARN when they reach the STARTING status.

newAttemptContainerDetail :: AttemptContainerDetail

Create a value of AttemptContainerDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logStreamName:AttemptContainerDetail', attemptContainerDetail_logStreamName - The name of the CloudWatch Logs log stream associated with the container. The log group for Batch jobs is /aws/batch/job. Each container attempt receives a log stream name when they reach the RUNNING status.

$sel:containerInstanceArn:AttemptContainerDetail', attemptContainerDetail_containerInstanceArn - The Amazon Resource Name (ARN) of the Amazon ECS container instance that hosts the job attempt.

$sel:exitCode:AttemptContainerDetail', attemptContainerDetail_exitCode - The exit code for the job attempt. A non-zero exit code is considered a failure.

$sel:reason:AttemptContainerDetail', attemptContainerDetail_reason - A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

$sel:networkInterfaces:AttemptContainerDetail', attemptContainerDetail_networkInterfaces - The network interfaces associated with the job attempt.

$sel:taskArn:AttemptContainerDetail', attemptContainerDetail_taskArn - The Amazon Resource Name (ARN) of the Amazon ECS task that's associated with the job attempt. Each container attempt receives a task ARN when they reach the STARTING status.

attemptContainerDetail_logStreamName :: Lens' AttemptContainerDetail (Maybe Text)

The name of the CloudWatch Logs log stream associated with the container. The log group for Batch jobs is /aws/batch/job. Each container attempt receives a log stream name when they reach the RUNNING status.

attemptContainerDetail_containerInstanceArn :: Lens' AttemptContainerDetail (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon ECS container instance that hosts the job attempt.

attemptContainerDetail_exitCode :: Lens' AttemptContainerDetail (Maybe Int)

The exit code for the job attempt. A non-zero exit code is considered a failure.

attemptContainerDetail_reason :: Lens' AttemptContainerDetail (Maybe Text)

A short (255 max characters) human-readable string to provide additional details about a running or stopped container.

attemptContainerDetail_networkInterfaces :: Lens' AttemptContainerDetail (Maybe [NetworkInterface])

The network interfaces associated with the job attempt.

attemptContainerDetail_taskArn :: Lens' AttemptContainerDetail (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon ECS task that's associated with the job attempt. Each container attempt receives a task ARN when they reach the STARTING status.