Network.AWS.Batch.Types.ComputeEnvironmentDetail

Description

Documentation

data ComputeEnvironmentDetail

An object representing an Batch compute environment.

See: newComputeEnvironmentDetail smart constructor.

Constructors

ComputeEnvironmentDetail'

Fields

status :: Maybe CEStatus

The current status of the compute environment (for example, CREATING or VALID).

serviceRole :: Maybe Text

The service role associated with the compute environment that allows Batch to make calls to Amazon Web Services API operations on your behalf. For more information, see Batch service IAM role in the Batch User Guide.

state :: Maybe CEState

The state of the compute environment. The valid values are ENABLED or DISABLED.

If the state is ENABLED, then the Batch scheduler can attempt to place jobs from an associated job queue on the compute resources within the environment. If the compute environment is managed, then it can scale its instances out or in automatically, based on the job queue demand.

If the state is DISABLED, then the Batch scheduler doesn't attempt to place jobs within the environment. Jobs in a STARTING or RUNNING state continue to progress normally. Managed compute environments in the DISABLED state don't scale out. However, they scale in to minvCpus value after instances become idle.

computeResources :: Maybe ComputeResource

The compute resources defined for the compute environment. For more information, see Compute Environments in the Batch User Guide.

tags :: Maybe (HashMap Text Text)

The tags applied to the compute environment.

type' :: Maybe CEType

The type of the compute environment: MANAGED or UNMANAGED. For more information, see Compute Environments in the Batch User Guide.

statusReason :: Maybe Text

A short, human-readable string to provide additional details about the current status of the compute environment.

computeEnvironmentName :: Text

The name of the compute environment. Up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

computeEnvironmentArn :: Text

The Amazon Resource Name (ARN) of the compute environment.

ecsClusterArn :: Text

The Amazon Resource Name (ARN) of the underlying Amazon ECS cluster used by the compute environment.

newComputeEnvironmentDetail

Arguments

:: Text
$sel:computeEnvironmentName:ComputeEnvironmentDetail'

-> Text
$sel:computeEnvironmentArn:ComputeEnvironmentDetail'

-> Text
$sel:ecsClusterArn:ComputeEnvironmentDetail'

-> ComputeEnvironmentDetail

Create a value of ComputeEnvironmentDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ComputeEnvironmentDetail', computeEnvironmentDetail_status - The current status of the compute environment (for example, CREATING or VALID).

$sel:serviceRole:ComputeEnvironmentDetail', computeEnvironmentDetail_serviceRole - The service role associated with the compute environment that allows Batch to make calls to Amazon Web Services API operations on your behalf. For more information, see Batch service IAM role in the Batch User Guide.

$sel:state:ComputeEnvironmentDetail', computeEnvironmentDetail_state - The state of the compute environment. The valid values are ENABLED or DISABLED.

If the state is ENABLED, then the Batch scheduler can attempt to place jobs from an associated job queue on the compute resources within the environment. If the compute environment is managed, then it can scale its instances out or in automatically, based on the job queue demand.

If the state is DISABLED, then the Batch scheduler doesn't attempt to place jobs within the environment. Jobs in a STARTING or RUNNING state continue to progress normally. Managed compute environments in the DISABLED state don't scale out. However, they scale in to minvCpus value after instances become idle.

$sel:computeResources:ComputeEnvironmentDetail', computeEnvironmentDetail_computeResources - The compute resources defined for the compute environment. For more information, see Compute Environments in the Batch User Guide.

$sel:tags:ComputeEnvironmentDetail', computeEnvironmentDetail_tags - The tags applied to the compute environment.

$sel:type':ComputeEnvironmentDetail', computeEnvironmentDetail_type - The type of the compute environment: MANAGED or UNMANAGED. For more information, see Compute Environments in the Batch User Guide.

$sel:statusReason:ComputeEnvironmentDetail', computeEnvironmentDetail_statusReason - A short, human-readable string to provide additional details about the current status of the compute environment.

$sel:computeEnvironmentName:ComputeEnvironmentDetail', computeEnvironmentDetail_computeEnvironmentName - The name of the compute environment. Up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

$sel:computeEnvironmentArn:ComputeEnvironmentDetail', computeEnvironmentDetail_computeEnvironmentArn - The Amazon Resource Name (ARN) of the compute environment.

$sel:ecsClusterArn:ComputeEnvironmentDetail', computeEnvironmentDetail_ecsClusterArn - The Amazon Resource Name (ARN) of the underlying Amazon ECS cluster used by the compute environment.

computeEnvironmentDetail_status :: Lens' ComputeEnvironmentDetail (Maybe CEStatus)

The current status of the compute environment (for example, CREATING or VALID).

computeEnvironmentDetail_serviceRole :: Lens' ComputeEnvironmentDetail (Maybe Text)

The service role associated with the compute environment that allows Batch to make calls to Amazon Web Services API operations on your behalf. For more information, see Batch service IAM role in the Batch User Guide.

computeEnvironmentDetail_state :: Lens' ComputeEnvironmentDetail (Maybe CEState)

The state of the compute environment. The valid values are ENABLED or DISABLED.

If the state is ENABLED, then the Batch scheduler can attempt to place jobs from an associated job queue on the compute resources within the environment. If the compute environment is managed, then it can scale its instances out or in automatically, based on the job queue demand.

If the state is DISABLED, then the Batch scheduler doesn't attempt to place jobs within the environment. Jobs in a STARTING or RUNNING state continue to progress normally. Managed compute environments in the DISABLED state don't scale out. However, they scale in to minvCpus value after instances become idle.

computeEnvironmentDetail_computeResources :: Lens' ComputeEnvironmentDetail (Maybe ComputeResource)

The compute resources defined for the compute environment. For more information, see Compute Environments in the Batch User Guide.

computeEnvironmentDetail_tags :: Lens' ComputeEnvironmentDetail (Maybe (HashMap Text Text))

The tags applied to the compute environment.

computeEnvironmentDetail_type :: Lens' ComputeEnvironmentDetail (Maybe CEType)

The type of the compute environment: MANAGED or UNMANAGED. For more information, see Compute Environments in the Batch User Guide.

computeEnvironmentDetail_statusReason :: Lens' ComputeEnvironmentDetail (Maybe Text)

A short, human-readable string to provide additional details about the current status of the compute environment.

computeEnvironmentDetail_computeEnvironmentName :: Lens' ComputeEnvironmentDetail Text

The name of the compute environment. Up to 128 letters (uppercase and lowercase), numbers, hyphens, and underscores are allowed.

computeEnvironmentDetail_computeEnvironmentArn :: Lens' ComputeEnvironmentDetail Text

The Amazon Resource Name (ARN) of the compute environment.

computeEnvironmentDetail_ecsClusterArn :: Lens' ComputeEnvironmentDetail Text

The Amazon Resource Name (ARN) of the underlying Amazon ECS cluster used by the compute environment.