Network.AWS.Batch.Types.ComputeResourceUpdate

Description

Documentation

data ComputeResourceUpdate

An object representing the attributes of a compute environment that can be updated. For more information, see Compute Environments in the Batch User Guide.

See: newComputeResourceUpdate smart constructor.

Constructors

ComputeResourceUpdate'

Fields

securityGroupIds :: Maybe [Text]

The Amazon EC2 security groups associated with instances launched in the compute environment. This parameter is required for Fargate compute resources, where it can contain up to 5 security groups. This can't be specified for EC2 compute resources. Providing an empty list is handled as if this parameter wasn't specified and no change is made.

minvCpus :: Maybe Int

The minimum number of Amazon EC2 vCPUs that an environment should maintain.

This parameter isn't applicable to jobs that are running on Fargate resources, and shouldn't be specified.

maxvCpus :: Maybe Int

The maximum number of Amazon EC2 vCPUs that an environment can reach.

With both BEST_FIT_PROGRESSIVE and SPOT_CAPACITY_OPTIMIZED allocation strategies, Batch might need to exceed maxvCpus to meet your capacity requirements. In this event, Batch never exceeds maxvCpus by more than a single instance. That is, no more than a single instance from among those specified in your compute environment.

desiredvCpus :: Maybe Int

The desired number of Amazon EC2 vCPUS in the compute environment.

This parameter isn't applicable to jobs that are running on Fargate resources, and shouldn't be specified.

subnets :: Maybe [Text]

The VPC subnets where the compute resources are launched. Fargate compute resources can contain up to 16 subnets. Providing an empty list will be handled as if this parameter wasn't specified and no change is made. This can't be specified for EC2 compute resources. For more information, see VPCs and Subnets in the Amazon VPC User Guide.

newComputeResourceUpdate :: ComputeResourceUpdate

Create a value of ComputeResourceUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:ComputeResourceUpdate', computeResourceUpdate_securityGroupIds - The Amazon EC2 security groups associated with instances launched in the compute environment. This parameter is required for Fargate compute resources, where it can contain up to 5 security groups. This can't be specified for EC2 compute resources. Providing an empty list is handled as if this parameter wasn't specified and no change is made.

$sel:minvCpus:ComputeResourceUpdate', computeResourceUpdate_minvCpus - The minimum number of Amazon EC2 vCPUs that an environment should maintain.

This parameter isn't applicable to jobs that are running on Fargate resources, and shouldn't be specified.

$sel:maxvCpus:ComputeResourceUpdate', computeResourceUpdate_maxvCpus - The maximum number of Amazon EC2 vCPUs that an environment can reach.

With both BEST_FIT_PROGRESSIVE and SPOT_CAPACITY_OPTIMIZED allocation strategies, Batch might need to exceed maxvCpus to meet your capacity requirements. In this event, Batch never exceeds maxvCpus by more than a single instance. That is, no more than a single instance from among those specified in your compute environment.

$sel:desiredvCpus:ComputeResourceUpdate', computeResourceUpdate_desiredvCpus - The desired number of Amazon EC2 vCPUS in the compute environment.

This parameter isn't applicable to jobs that are running on Fargate resources, and shouldn't be specified.

$sel:subnets:ComputeResourceUpdate', computeResourceUpdate_subnets - The VPC subnets where the compute resources are launched. Fargate compute resources can contain up to 16 subnets. Providing an empty list will be handled as if this parameter wasn't specified and no change is made. This can't be specified for EC2 compute resources. For more information, see VPCs and Subnets in the Amazon VPC User Guide.

computeResourceUpdate_securityGroupIds :: Lens' ComputeResourceUpdate (Maybe [Text])

The Amazon EC2 security groups associated with instances launched in the compute environment. This parameter is required for Fargate compute resources, where it can contain up to 5 security groups. This can't be specified for EC2 compute resources. Providing an empty list is handled as if this parameter wasn't specified and no change is made.

computeResourceUpdate_minvCpus :: Lens' ComputeResourceUpdate (Maybe Int)

The minimum number of Amazon EC2 vCPUs that an environment should maintain.

This parameter isn't applicable to jobs that are running on Fargate resources, and shouldn't be specified.

computeResourceUpdate_maxvCpus :: Lens' ComputeResourceUpdate (Maybe Int)

The maximum number of Amazon EC2 vCPUs that an environment can reach.

With both BEST_FIT_PROGRESSIVE and SPOT_CAPACITY_OPTIMIZED allocation strategies, Batch might need to exceed maxvCpus to meet your capacity requirements. In this event, Batch never exceeds maxvCpus by more than a single instance. That is, no more than a single instance from among those specified in your compute environment.

computeResourceUpdate_desiredvCpus :: Lens' ComputeResourceUpdate (Maybe Int)

The desired number of Amazon EC2 vCPUS in the compute environment.

This parameter isn't applicable to jobs that are running on Fargate resources, and shouldn't be specified.

computeResourceUpdate_subnets :: Lens' ComputeResourceUpdate (Maybe [Text])

The VPC subnets where the compute resources are launched. Fargate compute resources can contain up to 16 subnets. Providing an empty list will be handled as if this parameter wasn't specified and no change is made. This can't be specified for EC2 compute resources. For more information, see VPCs and Subnets in the Amazon VPC User Guide.