Network.AWS.Batch.Types.JobDefinition

Description

Documentation

data JobDefinition

An object representing an Batch job definition.

See: newJobDefinition smart constructor.

Constructors

JobDefinition'

Fields

status :: Maybe Text

The status of the job definition.

platformCapabilities :: Maybe [PlatformCapability]

The platform capabilities required by the job definition. If no value is specified, it defaults to EC2. Jobs run on Fargate resources specify FARGATE.

timeout :: Maybe JobTimeout

The timeout configuration for jobs that are submitted with this job definition. You can specify a timeout duration after which Batch terminates your jobs if they haven't finished.

nodeProperties :: Maybe NodeProperties

An object with various properties specific to multi-node parallel jobs.

If the job runs on Fargate resources, then you must not specify nodeProperties; use containerProperties instead.

tags :: Maybe (HashMap Text Text)

The tags applied to the job definition.

containerProperties :: Maybe ContainerProperties

An object with various properties specific to container-based jobs.

retryStrategy :: Maybe RetryStrategy

The retry strategy to use for failed jobs that are submitted with this job definition.

parameters :: Maybe (HashMap Text Text)

Default parameters or parameter substitution placeholders that are set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition. For more information about specifying parameters, see Job Definition Parameters in the Batch User Guide.

propagateTags :: Maybe Bool

Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags aren't propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state.

jobDefinitionName :: Text

The name of the job definition.

jobDefinitionArn :: Text

The Amazon Resource Name (ARN) for the job definition.

revision :: Int

The revision of the job definition.

type' :: Text

The type of job definition. If the job is run on Fargate resources, then multinode isn't supported. For more information about multi-node parallel jobs, see Creating a multi-node parallel job definition in the Batch User Guide.

newJobDefinition

Arguments

:: Text
$sel:jobDefinitionName:JobDefinition'

-> Text
$sel:jobDefinitionArn:JobDefinition'

-> Int
$sel:revision:JobDefinition'

-> Text
$sel:type':JobDefinition'

-> JobDefinition

Create a value of JobDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:JobDefinition', jobDefinition_status - The status of the job definition.

$sel:platformCapabilities:JobDefinition', jobDefinition_platformCapabilities - The platform capabilities required by the job definition. If no value is specified, it defaults to EC2. Jobs run on Fargate resources specify FARGATE.

$sel:timeout:JobDefinition', jobDefinition_timeout - The timeout configuration for jobs that are submitted with this job definition. You can specify a timeout duration after which Batch terminates your jobs if they haven't finished.

$sel:nodeProperties:JobDefinition', jobDefinition_nodeProperties - An object with various properties specific to multi-node parallel jobs.

If the job runs on Fargate resources, then you must not specify nodeProperties; use containerProperties instead.

$sel:tags:JobDefinition', jobDefinition_tags - The tags applied to the job definition.

$sel:containerProperties:JobDefinition', jobDefinition_containerProperties - An object with various properties specific to container-based jobs.

$sel:retryStrategy:JobDefinition', jobDefinition_retryStrategy - The retry strategy to use for failed jobs that are submitted with this job definition.

$sel:parameters:JobDefinition', jobDefinition_parameters - Default parameters or parameter substitution placeholders that are set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition. For more information about specifying parameters, see Job Definition Parameters in the Batch User Guide.

$sel:propagateTags:JobDefinition', jobDefinition_propagateTags - Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags aren't propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state.

$sel:jobDefinitionName:JobDefinition', jobDefinition_jobDefinitionName - The name of the job definition.

$sel:jobDefinitionArn:JobDefinition', jobDefinition_jobDefinitionArn - The Amazon Resource Name (ARN) for the job definition.

$sel:revision:JobDefinition', jobDefinition_revision - The revision of the job definition.

$sel:type':JobDefinition', jobDefinition_type - The type of job definition. If the job is run on Fargate resources, then multinode isn't supported. For more information about multi-node parallel jobs, see Creating a multi-node parallel job definition in the Batch User Guide.

jobDefinition_status :: Lens' JobDefinition (Maybe Text)

The status of the job definition.

jobDefinition_platformCapabilities :: Lens' JobDefinition (Maybe [PlatformCapability])

The platform capabilities required by the job definition. If no value is specified, it defaults to EC2. Jobs run on Fargate resources specify FARGATE.

jobDefinition_timeout :: Lens' JobDefinition (Maybe JobTimeout)

The timeout configuration for jobs that are submitted with this job definition. You can specify a timeout duration after which Batch terminates your jobs if they haven't finished.

jobDefinition_nodeProperties :: Lens' JobDefinition (Maybe NodeProperties)

An object with various properties specific to multi-node parallel jobs.

If the job runs on Fargate resources, then you must not specify nodeProperties; use containerProperties instead.

jobDefinition_tags :: Lens' JobDefinition (Maybe (HashMap Text Text))

The tags applied to the job definition.

jobDefinition_containerProperties :: Lens' JobDefinition (Maybe ContainerProperties)

An object with various properties specific to container-based jobs.

jobDefinition_retryStrategy :: Lens' JobDefinition (Maybe RetryStrategy)

The retry strategy to use for failed jobs that are submitted with this job definition.

jobDefinition_parameters :: Lens' JobDefinition (Maybe (HashMap Text Text))

Default parameters or parameter substitution placeholders that are set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition. For more information about specifying parameters, see Job Definition Parameters in the Batch User Guide.

jobDefinition_propagateTags :: Lens' JobDefinition (Maybe Bool)

Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags aren't propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state.

jobDefinition_jobDefinitionName :: Lens' JobDefinition Text

The name of the job definition.

jobDefinition_jobDefinitionArn :: Lens' JobDefinition Text

The Amazon Resource Name (ARN) for the job definition.

jobDefinition_revision :: Lens' JobDefinition Int

The revision of the job definition.

jobDefinition_type :: Lens' JobDefinition Text

The type of job definition. If the job is run on Fargate resources, then multinode isn't supported. For more information about multi-node parallel jobs, see Creating a multi-node parallel job definition in the Batch User Guide.