Network.AWS.Batch.Types.JobDetail

Description

Documentation

data JobDetail

An object representing an Batch job.

See: newJobDetail smart constructor.

Constructors

JobDetail'

Fields

container :: Maybe ContainerDetail

An object representing the details of the container that's associated with the job.

dependsOn :: Maybe [JobDependency]

A list of job IDs that this job depends on.

startedAt :: Maybe Integer

The Unix timestamp (in milliseconds) for when the job was started (when the job transitioned from the STARTING state to the RUNNING state). This parameter isn't provided for child jobs of array jobs or multi-node parallel jobs.

platformCapabilities :: Maybe [PlatformCapability]

The platform capabilities required by the job definition. If no value is specified, it defaults to EC2. Jobs run on Fargate resources specify FARGATE.

timeout :: Maybe JobTimeout

The timeout configuration for the job.

arrayProperties :: Maybe ArrayPropertiesDetail

The array properties of the job, if it is an array job.

createdAt :: Maybe Integer

The Unix timestamp (in milliseconds) for when the job was created. For non-array jobs and parent array jobs, this is when the job entered the SUBMITTED state (at the time SubmitJob was called). For array child jobs, this is when the child job was spawned by its parent and entered the PENDING state.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the job.

nodeDetails :: Maybe NodeDetails

An object representing the details of a node that's associated with a multi-node parallel job.

stoppedAt :: Maybe Integer

The Unix timestamp (in milliseconds) for when the job was stopped (when the job transitioned from the RUNNING state to a terminal state, such as SUCCEEDED or FAILED).

nodeProperties :: Maybe NodeProperties

An object representing the node properties of a multi-node parallel job.

This isn't applicable to jobs that are running on Fargate resources.

tags :: Maybe (HashMap Text Text)

The tags applied to the job.

attempts :: Maybe [AttemptDetail]

A list of job attempts associated with this job.

retryStrategy :: Maybe RetryStrategy

The retry strategy to use for this job if an attempt fails.

statusReason :: Maybe Text

A short, human-readable string to provide additional details about the current status of the job.

parameters :: Maybe (HashMap Text Text)

Additional parameters passed to the job that replace parameter substitution placeholders or override any corresponding parameter defaults from the job definition.

propagateTags :: Maybe Bool

Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags aren't propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state.

jobName :: Text

The name of the job.

jobId :: Text

The ID for the job.

jobQueue :: Text

The Amazon Resource Name (ARN) of the job queue that the job is associated with.

status :: JobStatus

The current status for the job.

If your jobs don't progress to STARTING, see Jobs Stuck in RUNNABLE Status in the troubleshooting section of the Batch User Guide.

jobDefinition :: Text

The job definition that's used by this job.

newJobDetail

Arguments

:: Text
$sel:jobName:JobDetail'

-> Text
$sel:jobId:JobDetail'

-> Text
$sel:jobQueue:JobDetail'

-> JobStatus
$sel:status:JobDetail'

-> Text
$sel:jobDefinition:JobDetail'

-> JobDetail

Create a value of JobDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:container:JobDetail', jobDetail_container - An object representing the details of the container that's associated with the job.

$sel:dependsOn:JobDetail', jobDetail_dependsOn - A list of job IDs that this job depends on.

$sel:startedAt:JobDetail', jobDetail_startedAt - The Unix timestamp (in milliseconds) for when the job was started (when the job transitioned from the STARTING state to the RUNNING state). This parameter isn't provided for child jobs of array jobs or multi-node parallel jobs.

$sel:platformCapabilities:JobDetail', jobDetail_platformCapabilities - The platform capabilities required by the job definition. If no value is specified, it defaults to EC2. Jobs run on Fargate resources specify FARGATE.

$sel:timeout:JobDetail', jobDetail_timeout - The timeout configuration for the job.

$sel:arrayProperties:JobDetail', jobDetail_arrayProperties - The array properties of the job, if it is an array job.

$sel:createdAt:JobDetail', jobDetail_createdAt - The Unix timestamp (in milliseconds) for when the job was created. For non-array jobs and parent array jobs, this is when the job entered the SUBMITTED state (at the time SubmitJob was called). For array child jobs, this is when the child job was spawned by its parent and entered the PENDING state.

$sel:jobArn:JobDetail', jobDetail_jobArn - The Amazon Resource Name (ARN) of the job.

$sel:nodeDetails:JobDetail', jobDetail_nodeDetails - An object representing the details of a node that's associated with a multi-node parallel job.

$sel:stoppedAt:JobDetail', jobDetail_stoppedAt - The Unix timestamp (in milliseconds) for when the job was stopped (when the job transitioned from the RUNNING state to a terminal state, such as SUCCEEDED or FAILED).

$sel:nodeProperties:JobDetail', jobDetail_nodeProperties - An object representing the node properties of a multi-node parallel job.

This isn't applicable to jobs that are running on Fargate resources.

$sel:tags:JobDetail', jobDetail_tags - The tags applied to the job.

$sel:attempts:JobDetail', jobDetail_attempts - A list of job attempts associated with this job.

$sel:retryStrategy:JobDetail', jobDetail_retryStrategy - The retry strategy to use for this job if an attempt fails.

$sel:statusReason:JobDetail', jobDetail_statusReason - A short, human-readable string to provide additional details about the current status of the job.

$sel:parameters:JobDetail', jobDetail_parameters - Additional parameters passed to the job that replace parameter substitution placeholders or override any corresponding parameter defaults from the job definition.

$sel:propagateTags:JobDetail', jobDetail_propagateTags - Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags aren't propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state.

$sel:jobName:JobDetail', jobDetail_jobName - The name of the job.

$sel:jobId:JobDetail', jobDetail_jobId - The ID for the job.

$sel:jobQueue:JobDetail', jobDetail_jobQueue - The Amazon Resource Name (ARN) of the job queue that the job is associated with.

$sel:status:JobDetail', jobDetail_status - The current status for the job.

If your jobs don't progress to STARTING, see Jobs Stuck in RUNNABLE Status in the troubleshooting section of the Batch User Guide.

$sel:jobDefinition:JobDetail', jobDetail_jobDefinition - The job definition that's used by this job.

jobDetail_container :: Lens' JobDetail (Maybe ContainerDetail)

An object representing the details of the container that's associated with the job.

jobDetail_dependsOn :: Lens' JobDetail (Maybe [JobDependency])

A list of job IDs that this job depends on.

jobDetail_startedAt :: Lens' JobDetail (Maybe Integer)

The Unix timestamp (in milliseconds) for when the job was started (when the job transitioned from the STARTING state to the RUNNING state). This parameter isn't provided for child jobs of array jobs or multi-node parallel jobs.

jobDetail_platformCapabilities :: Lens' JobDetail (Maybe [PlatformCapability])

The platform capabilities required by the job definition. If no value is specified, it defaults to EC2. Jobs run on Fargate resources specify FARGATE.

jobDetail_timeout :: Lens' JobDetail (Maybe JobTimeout)

The timeout configuration for the job.

jobDetail_arrayProperties :: Lens' JobDetail (Maybe ArrayPropertiesDetail)

The array properties of the job, if it is an array job.

jobDetail_createdAt :: Lens' JobDetail (Maybe Integer)

The Unix timestamp (in milliseconds) for when the job was created. For non-array jobs and parent array jobs, this is when the job entered the SUBMITTED state (at the time SubmitJob was called). For array child jobs, this is when the child job was spawned by its parent and entered the PENDING state.

jobDetail_jobArn :: Lens' JobDetail (Maybe Text)

The Amazon Resource Name (ARN) of the job.

jobDetail_nodeDetails :: Lens' JobDetail (Maybe NodeDetails)

An object representing the details of a node that's associated with a multi-node parallel job.

jobDetail_stoppedAt :: Lens' JobDetail (Maybe Integer)

The Unix timestamp (in milliseconds) for when the job was stopped (when the job transitioned from the RUNNING state to a terminal state, such as SUCCEEDED or FAILED).

jobDetail_nodeProperties :: Lens' JobDetail (Maybe NodeProperties)

An object representing the node properties of a multi-node parallel job.

This isn't applicable to jobs that are running on Fargate resources.

jobDetail_tags :: Lens' JobDetail (Maybe (HashMap Text Text))

The tags applied to the job.

jobDetail_attempts :: Lens' JobDetail (Maybe [AttemptDetail])

A list of job attempts associated with this job.

jobDetail_retryStrategy :: Lens' JobDetail (Maybe RetryStrategy)

The retry strategy to use for this job if an attempt fails.

jobDetail_statusReason :: Lens' JobDetail (Maybe Text)

A short, human-readable string to provide additional details about the current status of the job.

jobDetail_parameters :: Lens' JobDetail (Maybe (HashMap Text Text))

Additional parameters passed to the job that replace parameter substitution placeholders or override any corresponding parameter defaults from the job definition.

jobDetail_propagateTags :: Lens' JobDetail (Maybe Bool)

Specifies whether to propagate the tags from the job or job definition to the corresponding Amazon ECS task. If no value is specified, the tags aren't propagated. Tags can only be propagated to the tasks during task creation. For tags with the same name, job tags are given priority over job definitions tags. If the total number of combined tags from the job and job definition is over 50, the job is moved to the FAILED state.

jobDetail_jobName :: Lens' JobDetail Text

The name of the job.

jobDetail_jobId :: Lens' JobDetail Text

The ID for the job.

jobDetail_jobQueue :: Lens' JobDetail Text

The Amazon Resource Name (ARN) of the job queue that the job is associated with.

jobDetail_status :: Lens' JobDetail JobStatus

The current status for the job.

If your jobs don't progress to STARTING, see Jobs Stuck in RUNNABLE Status in the troubleshooting section of the Batch User Guide.

jobDetail_jobDefinition :: Lens' JobDetail Text

The job definition that's used by this job.