Network.AWS.Batch.Types.JobQueueDetail

Description

Documentation

data JobQueueDetail

An object representing the details of an Batch job queue.

See: newJobQueueDetail smart constructor.

Constructors

JobQueueDetail'

Fields

status :: Maybe JQStatus

The status of the job queue (for example, CREATING or VALID).

tags :: Maybe (HashMap Text Text)

The tags applied to the job queue. For more information, see Tagging your Batch resources in Batch User Guide.

statusReason :: Maybe Text

A short, human-readable string to provide additional details about the current status of the job queue.

jobQueueName :: Text

The name of the job queue.

jobQueueArn :: Text

The Amazon Resource Name (ARN) of the job queue.

state :: JQState

Describes the ability of the queue to accept new jobs. If the job queue state is ENABLED, it's able to accept jobs. If the job queue state is DISABLED, new jobs can't be added to the queue, but jobs already in the queue can finish.

priority :: Int

The priority of the job queue. Job queues with a higher priority (or a higher integer value for the priority parameter) are evaluated first when associated with the same compute environment. Priority is determined in descending order, for example, a job queue with a priority value of 10 is given scheduling preference over a job queue with a priority value of 1. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT); EC2 and Fargate compute environments can't be mixed.

computeEnvironmentOrder :: [ComputeEnvironmentOrder]

The compute environments that are attached to the job queue and the order that job placement is preferred. Compute environments are selected for job placement in ascending order.

newJobQueueDetail

Arguments

:: Text
$sel:jobQueueName:JobQueueDetail'

-> Text
$sel:jobQueueArn:JobQueueDetail'

-> JQState
$sel:state:JobQueueDetail'

-> Int
$sel:priority:JobQueueDetail'

-> JobQueueDetail

Create a value of JobQueueDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:JobQueueDetail', jobQueueDetail_status - The status of the job queue (for example, CREATING or VALID).

$sel:tags:JobQueueDetail', jobQueueDetail_tags - The tags applied to the job queue. For more information, see Tagging your Batch resources in Batch User Guide.

$sel:statusReason:JobQueueDetail', jobQueueDetail_statusReason - A short, human-readable string to provide additional details about the current status of the job queue.

$sel:jobQueueName:JobQueueDetail', jobQueueDetail_jobQueueName - The name of the job queue.

$sel:jobQueueArn:JobQueueDetail', jobQueueDetail_jobQueueArn - The Amazon Resource Name (ARN) of the job queue.

$sel:state:JobQueueDetail', jobQueueDetail_state - Describes the ability of the queue to accept new jobs. If the job queue state is ENABLED, it's able to accept jobs. If the job queue state is DISABLED, new jobs can't be added to the queue, but jobs already in the queue can finish.

$sel:priority:JobQueueDetail', jobQueueDetail_priority - The priority of the job queue. Job queues with a higher priority (or a higher integer value for the priority parameter) are evaluated first when associated with the same compute environment. Priority is determined in descending order, for example, a job queue with a priority value of 10 is given scheduling preference over a job queue with a priority value of 1. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT); EC2 and Fargate compute environments can't be mixed.

$sel:computeEnvironmentOrder:JobQueueDetail', jobQueueDetail_computeEnvironmentOrder - The compute environments that are attached to the job queue and the order that job placement is preferred. Compute environments are selected for job placement in ascending order.

jobQueueDetail_status :: Lens' JobQueueDetail (Maybe JQStatus)

The status of the job queue (for example, CREATING or VALID).

jobQueueDetail_tags :: Lens' JobQueueDetail (Maybe (HashMap Text Text))

The tags applied to the job queue. For more information, see Tagging your Batch resources in Batch User Guide.

jobQueueDetail_statusReason :: Lens' JobQueueDetail (Maybe Text)

A short, human-readable string to provide additional details about the current status of the job queue.

jobQueueDetail_jobQueueName :: Lens' JobQueueDetail Text

The name of the job queue.

jobQueueDetail_jobQueueArn :: Lens' JobQueueDetail Text

The Amazon Resource Name (ARN) of the job queue.

jobQueueDetail_state :: Lens' JobQueueDetail JQState

Describes the ability of the queue to accept new jobs. If the job queue state is ENABLED, it's able to accept jobs. If the job queue state is DISABLED, new jobs can't be added to the queue, but jobs already in the queue can finish.

jobQueueDetail_priority :: Lens' JobQueueDetail Int

The priority of the job queue. Job queues with a higher priority (or a higher integer value for the priority parameter) are evaluated first when associated with the same compute environment. Priority is determined in descending order, for example, a job queue with a priority value of 10 is given scheduling preference over a job queue with a priority value of 1. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT); EC2 and Fargate compute environments can't be mixed.

jobQueueDetail_computeEnvironmentOrder :: Lens' JobQueueDetail [ComputeEnvironmentOrder]

The compute environments that are attached to the job queue and the order that job placement is preferred. Compute environments are selected for job placement in ascending order.