Network.AWS.Batch.Types.JobSummary

Description

Documentation

data JobSummary

An object representing summary details of a job.

See: newJobSummary smart constructor.

Constructors

JobSummary'

Fields

stoppedAt :: Maybe Integer

The Unix timestamp for when the job was stopped (when the job transitioned from the RUNNING state to a terminal state, such as SUCCEEDED or FAILED).

status :: Maybe JobStatus

The current status for the job.

jobArn :: Maybe Text

The Amazon Resource Name (ARN) of the job.

createdAt :: Maybe Integer

The Unix timestamp for when the job was created. For non-array jobs and parent array jobs, this is when the job entered the SUBMITTED state (at the time SubmitJob was called). For array child jobs, this is when the child job was spawned by its parent and entered the PENDING state.

startedAt :: Maybe Integer

The Unix timestamp for when the job was started (when the job transitioned from the STARTING state to the RUNNING state).

container :: Maybe ContainerSummary

An object representing the details of the container that's associated with the job.

jobDefinition :: Maybe Text

The Amazon Resource Name (ARN) of the job definition.

statusReason :: Maybe Text

A short, human-readable string to provide additional details about the current status of the job.

arrayProperties :: Maybe ArrayPropertiesSummary

The array properties of the job, if it is an array job.

nodeProperties :: Maybe NodePropertiesSummary

The node properties for a single node in a job summary list.

This isn't applicable to jobs that are running on Fargate resources.

jobId :: Text

The ID of the job.

jobName :: Text

The name of the job.

newJobSummary

Arguments

:: Text
$sel:jobId:JobSummary'

-> Text
$sel:jobName:JobSummary'

-> JobSummary

Create a value of JobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stoppedAt:JobSummary', jobSummary_stoppedAt - The Unix timestamp for when the job was stopped (when the job transitioned from the RUNNING state to a terminal state, such as SUCCEEDED or FAILED).

$sel:status:JobSummary', jobSummary_status - The current status for the job.

$sel:jobArn:JobSummary', jobSummary_jobArn - The Amazon Resource Name (ARN) of the job.

$sel:createdAt:JobSummary', jobSummary_createdAt - The Unix timestamp for when the job was created. For non-array jobs and parent array jobs, this is when the job entered the SUBMITTED state (at the time SubmitJob was called). For array child jobs, this is when the child job was spawned by its parent and entered the PENDING state.

$sel:startedAt:JobSummary', jobSummary_startedAt - The Unix timestamp for when the job was started (when the job transitioned from the STARTING state to the RUNNING state).

$sel:container:JobSummary', jobSummary_container - An object representing the details of the container that's associated with the job.

$sel:jobDefinition:JobSummary', jobSummary_jobDefinition - The Amazon Resource Name (ARN) of the job definition.

$sel:statusReason:JobSummary', jobSummary_statusReason - A short, human-readable string to provide additional details about the current status of the job.

$sel:arrayProperties:JobSummary', jobSummary_arrayProperties - The array properties of the job, if it is an array job.

$sel:nodeProperties:JobSummary', jobSummary_nodeProperties - The node properties for a single node in a job summary list.

This isn't applicable to jobs that are running on Fargate resources.

$sel:jobId:JobSummary', jobSummary_jobId - The ID of the job.

$sel:jobName:JobSummary', jobSummary_jobName - The name of the job.

jobSummary_stoppedAt :: Lens' JobSummary (Maybe Integer)

The Unix timestamp for when the job was stopped (when the job transitioned from the RUNNING state to a terminal state, such as SUCCEEDED or FAILED).

jobSummary_status :: Lens' JobSummary (Maybe JobStatus)

The current status for the job.

jobSummary_jobArn :: Lens' JobSummary (Maybe Text)

The Amazon Resource Name (ARN) of the job.

jobSummary_createdAt :: Lens' JobSummary (Maybe Integer)

The Unix timestamp for when the job was created. For non-array jobs and parent array jobs, this is when the job entered the SUBMITTED state (at the time SubmitJob was called). For array child jobs, this is when the child job was spawned by its parent and entered the PENDING state.

jobSummary_startedAt :: Lens' JobSummary (Maybe Integer)

The Unix timestamp for when the job was started (when the job transitioned from the STARTING state to the RUNNING state).

jobSummary_container :: Lens' JobSummary (Maybe ContainerSummary)

An object representing the details of the container that's associated with the job.

jobSummary_jobDefinition :: Lens' JobSummary (Maybe Text)

The Amazon Resource Name (ARN) of the job definition.

jobSummary_statusReason :: Lens' JobSummary (Maybe Text)

A short, human-readable string to provide additional details about the current status of the job.

jobSummary_arrayProperties :: Lens' JobSummary (Maybe ArrayPropertiesSummary)

The array properties of the job, if it is an array job.

jobSummary_nodeProperties :: Lens' JobSummary (Maybe NodePropertiesSummary)

The node properties for a single node in a job summary list.

This isn't applicable to jobs that are running on Fargate resources.

jobSummary_jobId :: Lens' JobSummary Text

The ID of the job.

jobSummary_jobName :: Lens' JobSummary Text

The name of the job.