Network.AWS.Batch.Types.Tmpfs

Description

Documentation

data Tmpfs

The container path, mount options, and size of the tmpfs mount.

This object isn't applicable to jobs that are running on Fargate resources.

See: newTmpfs smart constructor.

Constructors

Tmpfs'

Fields

mountOptions :: Maybe [Text]

The list of tmpfs volume mount options.

Valid values: "defaults" | "ro" | "rw" | "suid" | "nosuid" | "dev" | "nodev" | "exec" | "noexec" | "sync" | "async" | "dirsync" | "remount" | "mand" | "nomand" | "atime" | "noatime" | "diratime" | "nodiratime" | "bind" | "rbind" | "unbindable" | "runbindable" | "private" | "rprivate" | "shared" | "rshared" | "slave" | "rslave" | "relatime" | "norelatime" | "strictatime" | "nostrictatime" | "mode" | "uid" | "gid" | "nr_inodes" | "nr_blocks" | "mpol"

containerPath :: Text

The absolute file path in the container where the tmpfs volume is mounted.

size :: Int

The size (in MiB) of the tmpfs volume.

newTmpfs

Arguments

:: Text
$sel:containerPath:Tmpfs'

-> Int
$sel:size:Tmpfs'

-> Tmpfs

Create a value of Tmpfs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mountOptions:Tmpfs', tmpfs_mountOptions - The list of tmpfs volume mount options.

Valid values: "defaults" | "ro" | "rw" | "suid" | "nosuid" | "dev" | "nodev" | "exec" | "noexec" | "sync" | "async" | "dirsync" | "remount" | "mand" | "nomand" | "atime" | "noatime" | "diratime" | "nodiratime" | "bind" | "rbind" | "unbindable" | "runbindable" | "private" | "rprivate" | "shared" | "rshared" | "slave" | "rslave" | "relatime" | "norelatime" | "strictatime" | "nostrictatime" | "mode" | "uid" | "gid" | "nr_inodes" | "nr_blocks" | "mpol"

$sel:containerPath:Tmpfs', tmpfs_containerPath - The absolute file path in the container where the tmpfs volume is mounted.

$sel:size:Tmpfs', tmpfs_size - The size (in MiB) of the tmpfs volume.

tmpfs_mountOptions :: Lens' Tmpfs (Maybe [Text])

The list of tmpfs volume mount options.

Valid values: "defaults" | "ro" | "rw" | "suid" | "nosuid" | "dev" | "nodev" | "exec" | "noexec" | "sync" | "async" | "dirsync" | "remount" | "mand" | "nomand" | "atime" | "noatime" | "diratime" | "nodiratime" | "bind" | "rbind" | "unbindable" | "runbindable" | "private" | "rprivate" | "shared" | "rshared" | "slave" | "rslave" | "relatime" | "norelatime" | "strictatime" | "nostrictatime" | "mode" | "uid" | "gid" | "nr_inodes" | "nr_blocks" | "mpol"

tmpfs_containerPath :: Lens' Tmpfs Text

The absolute file path in the container where the tmpfs volume is mounted.

tmpfs_size :: Lens' Tmpfs Int

The size (in MiB) of the tmpfs volume.