Network.AWS.Batch.UpdateJobQueue

Description

Updates a job queue.

Creating a Request

data UpdateJobQueue

Contains the parameters for UpdateJobQueue.

See: newUpdateJobQueue smart constructor.

Constructors

UpdateJobQueue'

Fields

state :: Maybe JQState

Describes the queue's ability to accept new jobs. If the job queue state is ENABLED, it can accept jobs. If the job queue state is DISABLED, new jobs can't be added to the queue, but jobs already in the queue can finish.

priority :: Maybe Int

The priority of the job queue. Job queues with a higher priority (or a higher integer value for the priority parameter) are evaluated first when associated with the same compute environment. Priority is determined in descending order, for example, a job queue with a priority value of 10 is given scheduling preference over a job queue with a priority value of 1. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT). EC2 and Fargate compute environments can't be mixed.

computeEnvironmentOrder :: Maybe [ComputeEnvironmentOrder]

Details the set of compute environments mapped to a job queue and their order relative to each other. This is one of the parameters used by the job scheduler to determine which compute environment should run a given job. Compute environments must be in the VALID state before you can associate them with a job queue. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT). EC2 and Fargate compute environments can't be mixed.

All compute environments that are associated with a job queue must share the same architecture. Batch doesn't support mixing compute environment architecture types in a single job queue.

jobQueue :: Text

The name or the Amazon Resource Name (ARN) of the job queue.

newUpdateJobQueue

Arguments

:: Text
$sel:jobQueue:UpdateJobQueue'

-> UpdateJobQueue

Create a value of UpdateJobQueue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:UpdateJobQueue', updateJobQueue_state - Describes the queue's ability to accept new jobs. If the job queue state is ENABLED, it can accept jobs. If the job queue state is DISABLED, new jobs can't be added to the queue, but jobs already in the queue can finish.

$sel:priority:UpdateJobQueue', updateJobQueue_priority - The priority of the job queue. Job queues with a higher priority (or a higher integer value for the priority parameter) are evaluated first when associated with the same compute environment. Priority is determined in descending order, for example, a job queue with a priority value of 10 is given scheduling preference over a job queue with a priority value of 1. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT). EC2 and Fargate compute environments can't be mixed.

$sel:computeEnvironmentOrder:UpdateJobQueue', updateJobQueue_computeEnvironmentOrder - Details the set of compute environments mapped to a job queue and their order relative to each other. This is one of the parameters used by the job scheduler to determine which compute environment should run a given job. Compute environments must be in the VALID state before you can associate them with a job queue. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT). EC2 and Fargate compute environments can't be mixed.

All compute environments that are associated with a job queue must share the same architecture. Batch doesn't support mixing compute environment architecture types in a single job queue.

$sel:jobQueue:UpdateJobQueue', updateJobQueue_jobQueue - The name or the Amazon Resource Name (ARN) of the job queue.

Request Lenses

updateJobQueue_state :: Lens' UpdateJobQueue (Maybe JQState)

Describes the queue's ability to accept new jobs. If the job queue state is ENABLED, it can accept jobs. If the job queue state is DISABLED, new jobs can't be added to the queue, but jobs already in the queue can finish.

updateJobQueue_priority :: Lens' UpdateJobQueue (Maybe Int)

The priority of the job queue. Job queues with a higher priority (or a higher integer value for the priority parameter) are evaluated first when associated with the same compute environment. Priority is determined in descending order, for example, a job queue with a priority value of 10 is given scheduling preference over a job queue with a priority value of 1. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT). EC2 and Fargate compute environments can't be mixed.

updateJobQueue_computeEnvironmentOrder :: Lens' UpdateJobQueue (Maybe [ComputeEnvironmentOrder])

Details the set of compute environments mapped to a job queue and their order relative to each other. This is one of the parameters used by the job scheduler to determine which compute environment should run a given job. Compute environments must be in the VALID state before you can associate them with a job queue. All of the compute environments must be either EC2 (EC2 or SPOT) or Fargate (FARGATE or FARGATE_SPOT). EC2 and Fargate compute environments can't be mixed.

All compute environments that are associated with a job queue must share the same architecture. Batch doesn't support mixing compute environment architecture types in a single job queue.

updateJobQueue_jobQueue :: Lens' UpdateJobQueue Text

The name or the Amazon Resource Name (ARN) of the job queue.

Destructuring the Response

data UpdateJobQueueResponse

See: newUpdateJobQueueResponse smart constructor.

Constructors

UpdateJobQueueResponse'

Fields

jobQueueArn :: Maybe Text

The Amazon Resource Name (ARN) of the job queue.

jobQueueName :: Maybe Text

The name of the job queue.

httpStatus :: Int

The response's http status code.

newUpdateJobQueueResponse

Arguments

:: Int
$sel:httpStatus:UpdateJobQueueResponse'

-> UpdateJobQueueResponse

Create a value of UpdateJobQueueResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobQueueArn:UpdateJobQueueResponse', updateJobQueueResponse_jobQueueArn - The Amazon Resource Name (ARN) of the job queue.

$sel:jobQueueName:UpdateJobQueueResponse', updateJobQueueResponse_jobQueueName - The name of the job queue.

$sel:httpStatus:UpdateJobQueueResponse', updateJobQueueResponse_httpStatus - The response's http status code.

Response Lenses

updateJobQueueResponse_jobQueueArn :: Lens' UpdateJobQueueResponse (Maybe Text)

The Amazon Resource Name (ARN) of the job queue.

updateJobQueueResponse_jobQueueName :: Lens' UpdateJobQueueResponse (Maybe Text)

The name of the job queue.

updateJobQueueResponse_httpStatus :: Lens' UpdateJobQueueResponse Int

The response's http status code.