Network.AWS.Braket.CancelQuantumTask

Description

Cancels the specified task.

Creating a Request

data CancelQuantumTask

See: newCancelQuantumTask smart constructor.

Constructors

CancelQuantumTask'

Fields

clientToken :: Text

The client token associated with the request.

quantumTaskArn :: Text

The ARN of the task to cancel.

newCancelQuantumTask

Arguments

:: Text
$sel:clientToken:CancelQuantumTask'

-> Text
$sel:quantumTaskArn:CancelQuantumTask'

-> CancelQuantumTask

Create a value of CancelQuantumTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CancelQuantumTask', cancelQuantumTask_clientToken - The client token associated with the request.

$sel:quantumTaskArn:CancelQuantumTask', cancelQuantumTask_quantumTaskArn - The ARN of the task to cancel.

Request Lenses

cancelQuantumTask_clientToken :: Lens' CancelQuantumTask Text

The client token associated with the request.

cancelQuantumTask_quantumTaskArn :: Lens' CancelQuantumTask Text

The ARN of the task to cancel.

Destructuring the Response

data CancelQuantumTaskResponse

See: newCancelQuantumTaskResponse smart constructor.

Constructors

CancelQuantumTaskResponse'

Fields

httpStatus :: Int

The response's http status code.

cancellationStatus :: CancellationStatus

The status of the cancellation request.

quantumTaskArn :: Text

The ARN of the task.

newCancelQuantumTaskResponse

Arguments

:: Int
$sel:httpStatus:CancelQuantumTaskResponse'

-> CancellationStatus
$sel:cancellationStatus:CancelQuantumTaskResponse'

-> Text
$sel:quantumTaskArn:CancelQuantumTask'

-> CancelQuantumTaskResponse

Create a value of CancelQuantumTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CancelQuantumTaskResponse', cancelQuantumTaskResponse_httpStatus - The response's http status code.

$sel:cancellationStatus:CancelQuantumTaskResponse', cancelQuantumTaskResponse_cancellationStatus - The status of the cancellation request.

$sel:quantumTaskArn:CancelQuantumTask', cancelQuantumTaskResponse_quantumTaskArn - The ARN of the task.

Response Lenses

cancelQuantumTaskResponse_httpStatus :: Lens' CancelQuantumTaskResponse Int

The response's http status code.

cancelQuantumTaskResponse_cancellationStatus :: Lens' CancelQuantumTaskResponse CancellationStatus

The status of the cancellation request.

cancelQuantumTaskResponse_quantumTaskArn :: Lens' CancelQuantumTaskResponse Text

The ARN of the task.