Network.AWS.Braket.CreateQuantumTask

Description

Creates a quantum task.

Creating a Request

data CreateQuantumTask

See: newCreateQuantumTask smart constructor.

Constructors

CreateQuantumTask'

Fields

deviceParameters :: Maybe Text

The parameters for the device to run the task on.

tags :: Maybe (HashMap Text Text)

Tags to be added to the quantum task you're creating.

action :: Text

The action associated with the task.

clientToken :: Text

The client token associated with the request.

deviceArn :: Text

The ARN of the device to run the task on.

outputS3Bucket :: Text

The S3 bucket to store task result files in.

outputS3KeyPrefix :: Text

The key prefix for the location in the S3 bucket to store task results in.

shots :: Natural

The number of shots to use for the task.

newCreateQuantumTask

Arguments

:: Text
$sel:action:CreateQuantumTask'

-> Text
$sel:clientToken:CreateQuantumTask'

-> Text
$sel:deviceArn:CreateQuantumTask'

-> Text
$sel:outputS3Bucket:CreateQuantumTask'

-> Text
$sel:outputS3KeyPrefix:CreateQuantumTask'

-> Natural
$sel:shots:CreateQuantumTask'

-> CreateQuantumTask

Create a value of CreateQuantumTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceParameters:CreateQuantumTask', createQuantumTask_deviceParameters - The parameters for the device to run the task on.

$sel:tags:CreateQuantumTask', createQuantumTask_tags - Tags to be added to the quantum task you're creating.

$sel:action:CreateQuantumTask', createQuantumTask_action - The action associated with the task.

$sel:clientToken:CreateQuantumTask', createQuantumTask_clientToken - The client token associated with the request.

$sel:deviceArn:CreateQuantumTask', createQuantumTask_deviceArn - The ARN of the device to run the task on.

$sel:outputS3Bucket:CreateQuantumTask', createQuantumTask_outputS3Bucket - The S3 bucket to store task result files in.

$sel:outputS3KeyPrefix:CreateQuantumTask', createQuantumTask_outputS3KeyPrefix - The key prefix for the location in the S3 bucket to store task results in.

$sel:shots:CreateQuantumTask', createQuantumTask_shots - The number of shots to use for the task.

Request Lenses

createQuantumTask_deviceParameters :: Lens' CreateQuantumTask (Maybe Text)

The parameters for the device to run the task on.

createQuantumTask_tags :: Lens' CreateQuantumTask (Maybe (HashMap Text Text))

Tags to be added to the quantum task you're creating.

createQuantumTask_action :: Lens' CreateQuantumTask Text

The action associated with the task.

createQuantumTask_clientToken :: Lens' CreateQuantumTask Text

The client token associated with the request.

createQuantumTask_deviceArn :: Lens' CreateQuantumTask Text

The ARN of the device to run the task on.

createQuantumTask_outputS3Bucket :: Lens' CreateQuantumTask Text

The S3 bucket to store task result files in.

createQuantumTask_outputS3KeyPrefix :: Lens' CreateQuantumTask Text

The key prefix for the location in the S3 bucket to store task results in.

createQuantumTask_shots :: Lens' CreateQuantumTask Natural

The number of shots to use for the task.

Destructuring the Response

data CreateQuantumTaskResponse

See: newCreateQuantumTaskResponse smart constructor.

Constructors

CreateQuantumTaskResponse'

Fields

httpStatus :: Int

The response's http status code.

quantumTaskArn :: Text

The ARN of the task created by the request.

newCreateQuantumTaskResponse

Arguments

:: Int
$sel:httpStatus:CreateQuantumTaskResponse'

-> Text
$sel:quantumTaskArn:CreateQuantumTaskResponse'

-> CreateQuantumTaskResponse

Create a value of CreateQuantumTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateQuantumTaskResponse', createQuantumTaskResponse_httpStatus - The response's http status code.

$sel:quantumTaskArn:CreateQuantumTaskResponse', createQuantumTaskResponse_quantumTaskArn - The ARN of the task created by the request.

Response Lenses

createQuantumTaskResponse_httpStatus :: Lens' CreateQuantumTaskResponse Int

The response's http status code.

createQuantumTaskResponse_quantumTaskArn :: Lens' CreateQuantumTaskResponse Text

The ARN of the task created by the request.