Network.AWS.Braket.Lens

Description

Operations

GetQuantumTask

getQuantumTask_quantumTaskArn :: Lens' GetQuantumTask Text

the ARN of the task to retrieve.

getQuantumTaskResponse_failureReason :: Lens' GetQuantumTaskResponse (Maybe Text)

The reason that a task failed.

getQuantumTaskResponse_endedAt :: Lens' GetQuantumTaskResponse (Maybe UTCTime)

The time at which the task ended.

getQuantumTaskResponse_tags :: Lens' GetQuantumTaskResponse (Maybe (HashMap Text Text))

The tags that belong to this task.

getQuantumTaskResponse_httpStatus :: Lens' GetQuantumTaskResponse Int

The response's http status code.

getQuantumTaskResponse_createdAt :: Lens' GetQuantumTaskResponse UTCTime

The time at which the task was created.

getQuantumTaskResponse_deviceArn :: Lens' GetQuantumTaskResponse Text

The ARN of the device the task was run on.

getQuantumTaskResponse_deviceParameters :: Lens' GetQuantumTaskResponse Text

The parameters for the device on which the task ran.

getQuantumTaskResponse_outputS3Bucket :: Lens' GetQuantumTaskResponse Text

The S3 bucket where task results are stored.

getQuantumTaskResponse_outputS3Directory :: Lens' GetQuantumTaskResponse Text

The folder in the S3 bucket where task results are stored.

getQuantumTaskResponse_quantumTaskArn :: Lens' GetQuantumTaskResponse Text

The ARN of the task.

getQuantumTaskResponse_shots :: Lens' GetQuantumTaskResponse Integer

The number of shots used in the task.

getQuantumTaskResponse_status :: Lens' GetQuantumTaskResponse QuantumTaskStatus

The status of the task.

SearchQuantumTasks

searchQuantumTasks_nextToken :: Lens' SearchQuantumTasks (Maybe Text)

A token used for pagination of results returned in the response. Use the token returned from the previous request continue results where the previous request ended.

searchQuantumTasks_maxResults :: Lens' SearchQuantumTasks (Maybe Natural)

Maximum number of results to return in the response.

searchQuantumTasks_filters :: Lens' SearchQuantumTasks [SearchQuantumTasksFilter]

Array of SearchQuantumTasksFilter objects.

searchQuantumTasksResponse_nextToken :: Lens' SearchQuantumTasksResponse (Maybe Text)

A token used for pagination of results, or null if there are no additional results. Use the token value in a subsequent request to continue results where the previous request ended.

searchQuantumTasksResponse_httpStatus :: Lens' SearchQuantumTasksResponse Int

The response's http status code.

searchQuantumTasksResponse_quantumTasks :: Lens' SearchQuantumTasksResponse [QuantumTaskSummary]

An array of QuantumTaskSummary objects for tasks that match the specified filters.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

Specify the resourceArn for the resource whose tags to display.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

Displays the key, value pairs of tags associated with this resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

GetDevice

getDevice_deviceArn :: Lens' GetDevice Text

The ARN of the device to retrieve.

getDeviceResponse_httpStatus :: Lens' GetDeviceResponse Int

The response's http status code.

getDeviceResponse_deviceArn :: Lens' GetDeviceResponse Text

The ARN of the device.

getDeviceResponse_deviceCapabilities :: Lens' GetDeviceResponse Text

Details about the capabilities of the device.

getDeviceResponse_deviceName :: Lens' GetDeviceResponse Text

The name of the device.

getDeviceResponse_deviceStatus :: Lens' GetDeviceResponse DeviceStatus

The status of the device.

getDeviceResponse_deviceType :: Lens' GetDeviceResponse DeviceType

The type of the device.

getDeviceResponse_providerName :: Lens' GetDeviceResponse Text

The name of the partner company for the device.

CancelQuantumTask

cancelQuantumTask_clientToken :: Lens' CancelQuantumTask Text

The client token associated with the request.

cancelQuantumTask_quantumTaskArn :: Lens' CancelQuantumTask Text

The ARN of the task to cancel.

cancelQuantumTaskResponse_httpStatus :: Lens' CancelQuantumTaskResponse Int

The response's http status code.

cancelQuantumTaskResponse_cancellationStatus :: Lens' CancelQuantumTaskResponse CancellationStatus

The status of the cancellation request.

cancelQuantumTaskResponse_quantumTaskArn :: Lens' CancelQuantumTaskResponse Text

The ARN of the task.

SearchDevices

searchDevices_nextToken :: Lens' SearchDevices (Maybe Text)

A token used for pagination of results returned in the response. Use the token returned from the previous request continue results where the previous request ended.

searchDevices_maxResults :: Lens' SearchDevices (Maybe Natural)

The maximum number of results to return in the response.

searchDevices_filters :: Lens' SearchDevices [SearchDevicesFilter]

The filter values to use to search for a device.

searchDevicesResponse_nextToken :: Lens' SearchDevicesResponse (Maybe Text)

A token used for pagination of results, or null if there are no additional results. Use the token value in a subsequent request to continue results where the previous request ended.

searchDevicesResponse_httpStatus :: Lens' SearchDevicesResponse Int

The response's http status code.

searchDevicesResponse_devices :: Lens' SearchDevicesResponse [DeviceSummary]

An array of DeviceSummary objects for devices that match the specified filter values.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

Specify the resourceArn of the resource to which a tag will be added.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

Specify the tags to add to the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

Specify the resourceArn for the resource from which to remove the tags.

untagResource_tagKeys :: Lens' UntagResource [Text]

Specify the keys for the tags to remove from the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

CreateQuantumTask

createQuantumTask_deviceParameters :: Lens' CreateQuantumTask (Maybe Text)

The parameters for the device to run the task on.

createQuantumTask_tags :: Lens' CreateQuantumTask (Maybe (HashMap Text Text))

Tags to be added to the quantum task you're creating.

createQuantumTask_action :: Lens' CreateQuantumTask Text

The action associated with the task.

createQuantumTask_clientToken :: Lens' CreateQuantumTask Text

The client token associated with the request.

createQuantumTask_deviceArn :: Lens' CreateQuantumTask Text

The ARN of the device to run the task on.

createQuantumTask_outputS3Bucket :: Lens' CreateQuantumTask Text

The S3 bucket to store task result files in.

createQuantumTask_outputS3KeyPrefix :: Lens' CreateQuantumTask Text

The key prefix for the location in the S3 bucket to store task results in.

createQuantumTask_shots :: Lens' CreateQuantumTask Natural

The number of shots to use for the task.

createQuantumTaskResponse_httpStatus :: Lens' CreateQuantumTaskResponse Int

The response's http status code.

createQuantumTaskResponse_quantumTaskArn :: Lens' CreateQuantumTaskResponse Text

The ARN of the task created by the request.

Types

DeviceSummary

deviceSummary_deviceArn :: Lens' DeviceSummary Text

The ARN of the device.

deviceSummary_deviceName :: Lens' DeviceSummary Text

The name of the device.

deviceSummary_deviceStatus :: Lens' DeviceSummary DeviceStatus

The status of the device.

deviceSummary_deviceType :: Lens' DeviceSummary DeviceType

The type of the device.

deviceSummary_providerName :: Lens' DeviceSummary Text

The provider of the device.

QuantumTaskSummary

quantumTaskSummary_endedAt :: Lens' QuantumTaskSummary (Maybe UTCTime)

The time at which the task finished.

quantumTaskSummary_tags :: Lens' QuantumTaskSummary (Maybe (HashMap Text Text))

Displays the key, value pairs of tags associated with this quantum task.

quantumTaskSummary_createdAt :: Lens' QuantumTaskSummary UTCTime

The time at which the task was created.

quantumTaskSummary_deviceArn :: Lens' QuantumTaskSummary Text

The ARN of the device the task ran on.

quantumTaskSummary_outputS3Bucket :: Lens' QuantumTaskSummary Text

The S3 bucket where the task result file is stored..

quantumTaskSummary_outputS3Directory :: Lens' QuantumTaskSummary Text

The folder in the S3 bucket where the task result file is stored.

quantumTaskSummary_quantumTaskArn :: Lens' QuantumTaskSummary Text

The ARN of the task.

quantumTaskSummary_shots :: Lens' QuantumTaskSummary Integer

The shots used for the task.

quantumTaskSummary_status :: Lens' QuantumTaskSummary QuantumTaskStatus

The status of the task.

SearchDevicesFilter

searchDevicesFilter_name :: Lens' SearchDevicesFilter Text

The name to use to filter results.

searchDevicesFilter_values :: Lens' SearchDevicesFilter (NonEmpty Text)

The values to use to filter results.

SearchQuantumTasksFilter

searchQuantumTasksFilter_name :: Lens' SearchQuantumTasksFilter Text

The name of the device used for the task.

searchQuantumTasksFilter_operator :: Lens' SearchQuantumTasksFilter SearchQuantumTasksFilterOperator

An operator to use in the filter.

searchQuantumTasksFilter_values :: Lens' SearchQuantumTasksFilter (NonEmpty Text)

The values to use for the filter.