Network.AWS.Braket.Types

Description

Service Configuration

defaultService :: Service

API version 2019-09-01 of the Amazon Braket SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The input fails to satisfy the constraints specified by an AWS service.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_DeviceOfflineException :: AsError a => Getting (First ServiceError) a ServiceError

The specified device is currently offline.

_DeviceRetiredException :: AsError a => Getting (First ServiceError) a ServiceError

The specified device has been retired.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

An error occurred due to a conflict.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request failed because a service quota is exceeded.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The throttling rate limit is met.

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing has failed because of an unknown error, exception, or failure.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource was not found.

CancellationStatus

newtype CancellationStatus

Constructors

CancellationStatus'

Fields

fromCancellationStatus :: Text

Bundled Patterns

pattern CancellationStatus_CANCELLED :: CancellationStatus

pattern CancellationStatus_CANCELLING :: CancellationStatus

DeviceStatus

newtype DeviceStatus

Constructors

DeviceStatus'

Fields

fromDeviceStatus :: Text

Bundled Patterns

pattern DeviceStatus_OFFLINE :: DeviceStatus

pattern DeviceStatus_ONLINE :: DeviceStatus

pattern DeviceStatus_RETIRED :: DeviceStatus

DeviceType

newtype DeviceType

Constructors

DeviceType'

Fields

fromDeviceType :: Text

Bundled Patterns

pattern DeviceType_QPU :: DeviceType

pattern DeviceType_SIMULATOR :: DeviceType

QuantumTaskStatus

newtype QuantumTaskStatus

Constructors

QuantumTaskStatus'

Fields

fromQuantumTaskStatus :: Text

Bundled Patterns

pattern QuantumTaskStatus_CANCELLED :: QuantumTaskStatus

pattern QuantumTaskStatus_CANCELLING :: QuantumTaskStatus

pattern QuantumTaskStatus_COMPLETED :: QuantumTaskStatus

pattern QuantumTaskStatus_CREATED :: QuantumTaskStatus

pattern QuantumTaskStatus_FAILED :: QuantumTaskStatus

pattern QuantumTaskStatus_QUEUED :: QuantumTaskStatus

pattern QuantumTaskStatus_RUNNING :: QuantumTaskStatus

SearchQuantumTasksFilterOperator

newtype SearchQuantumTasksFilterOperator

Constructors

SearchQuantumTasksFilterOperator'

Fields

fromSearchQuantumTasksFilterOperator :: Text

Bundled Patterns

pattern SearchQuantumTasksFilterOperator_BETWEEN :: SearchQuantumTasksFilterOperator

pattern SearchQuantumTasksFilterOperator_EQUAL :: SearchQuantumTasksFilterOperator

pattern SearchQuantumTasksFilterOperator_GT :: SearchQuantumTasksFilterOperator

pattern SearchQuantumTasksFilterOperator_GTE :: SearchQuantumTasksFilterOperator

pattern SearchQuantumTasksFilterOperator_LT :: SearchQuantumTasksFilterOperator

pattern SearchQuantumTasksFilterOperator_LTE :: SearchQuantumTasksFilterOperator

DeviceSummary

data DeviceSummary

Includes information about the device.

See: newDeviceSummary smart constructor.

Constructors

DeviceSummary'

Fields

deviceArn :: Text

The ARN of the device.

deviceName :: Text

The name of the device.

deviceStatus :: DeviceStatus

The status of the device.

deviceType :: DeviceType

The type of the device.

providerName :: Text

The provider of the device.

newDeviceSummary

Arguments

:: Text
$sel:deviceArn:DeviceSummary'

-> Text
$sel:deviceName:DeviceSummary'

-> DeviceStatus
$sel:deviceStatus:DeviceSummary'

-> DeviceType
$sel:deviceType:DeviceSummary'

-> Text
$sel:providerName:DeviceSummary'

-> DeviceSummary

Create a value of DeviceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceArn:DeviceSummary', deviceSummary_deviceArn - The ARN of the device.

$sel:deviceName:DeviceSummary', deviceSummary_deviceName - The name of the device.

$sel:deviceStatus:DeviceSummary', deviceSummary_deviceStatus - The status of the device.

$sel:deviceType:DeviceSummary', deviceSummary_deviceType - The type of the device.

$sel:providerName:DeviceSummary', deviceSummary_providerName - The provider of the device.

deviceSummary_deviceArn :: Lens' DeviceSummary Text

The ARN of the device.

deviceSummary_deviceName :: Lens' DeviceSummary Text

The name of the device.

deviceSummary_deviceStatus :: Lens' DeviceSummary DeviceStatus

The status of the device.

deviceSummary_deviceType :: Lens' DeviceSummary DeviceType

The type of the device.

deviceSummary_providerName :: Lens' DeviceSummary Text

The provider of the device.

QuantumTaskSummary

data QuantumTaskSummary

Includes information about a quantum task.

See: newQuantumTaskSummary smart constructor.

Constructors

QuantumTaskSummary'

Fields

endedAt :: Maybe POSIX

The time at which the task finished.

tags :: Maybe (HashMap Text Text)

Displays the key, value pairs of tags associated with this quantum task.

createdAt :: POSIX

The time at which the task was created.

deviceArn :: Text

The ARN of the device the task ran on.

outputS3Bucket :: Text

The S3 bucket where the task result file is stored..

outputS3Directory :: Text

The folder in the S3 bucket where the task result file is stored.

quantumTaskArn :: Text

The ARN of the task.

shots :: Integer

The shots used for the task.

status :: QuantumTaskStatus

The status of the task.

newQuantumTaskSummary

Arguments

:: UTCTime
$sel:createdAt:QuantumTaskSummary'

-> Text
$sel:deviceArn:QuantumTaskSummary'

-> Text
$sel:outputS3Bucket:QuantumTaskSummary'

-> Text
$sel:outputS3Directory:QuantumTaskSummary'

-> Text
$sel:quantumTaskArn:QuantumTaskSummary'

-> Integer
$sel:shots:QuantumTaskSummary'

-> QuantumTaskStatus
$sel:status:QuantumTaskSummary'

-> QuantumTaskSummary

Create a value of QuantumTaskSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endedAt:QuantumTaskSummary', quantumTaskSummary_endedAt - The time at which the task finished.

$sel:tags:QuantumTaskSummary', quantumTaskSummary_tags - Displays the key, value pairs of tags associated with this quantum task.

$sel:createdAt:QuantumTaskSummary', quantumTaskSummary_createdAt - The time at which the task was created.

$sel:deviceArn:QuantumTaskSummary', quantumTaskSummary_deviceArn - The ARN of the device the task ran on.

$sel:outputS3Bucket:QuantumTaskSummary', quantumTaskSummary_outputS3Bucket - The S3 bucket where the task result file is stored..

$sel:outputS3Directory:QuantumTaskSummary', quantumTaskSummary_outputS3Directory - The folder in the S3 bucket where the task result file is stored.

$sel:quantumTaskArn:QuantumTaskSummary', quantumTaskSummary_quantumTaskArn - The ARN of the task.

$sel:shots:QuantumTaskSummary', quantumTaskSummary_shots - The shots used for the task.

$sel:status:QuantumTaskSummary', quantumTaskSummary_status - The status of the task.

quantumTaskSummary_endedAt :: Lens' QuantumTaskSummary (Maybe UTCTime)

The time at which the task finished.

quantumTaskSummary_tags :: Lens' QuantumTaskSummary (Maybe (HashMap Text Text))

Displays the key, value pairs of tags associated with this quantum task.

quantumTaskSummary_createdAt :: Lens' QuantumTaskSummary UTCTime

The time at which the task was created.

quantumTaskSummary_deviceArn :: Lens' QuantumTaskSummary Text

The ARN of the device the task ran on.

quantumTaskSummary_outputS3Bucket :: Lens' QuantumTaskSummary Text

The S3 bucket where the task result file is stored..

quantumTaskSummary_outputS3Directory :: Lens' QuantumTaskSummary Text

The folder in the S3 bucket where the task result file is stored.

quantumTaskSummary_quantumTaskArn :: Lens' QuantumTaskSummary Text

The ARN of the task.

quantumTaskSummary_shots :: Lens' QuantumTaskSummary Integer

The shots used for the task.

quantumTaskSummary_status :: Lens' QuantumTaskSummary QuantumTaskStatus

The status of the task.

SearchDevicesFilter

data SearchDevicesFilter

The filter to use for searching devices.

See: newSearchDevicesFilter smart constructor.

Constructors

SearchDevicesFilter'

Fields

name :: Text

The name to use to filter results.

values :: NonEmpty Text

The values to use to filter results.

newSearchDevicesFilter

Arguments

:: Text
$sel:name:SearchDevicesFilter'

-> NonEmpty Text
$sel:values:SearchDevicesFilter'

-> SearchDevicesFilter

Create a value of SearchDevicesFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:SearchDevicesFilter', searchDevicesFilter_name - The name to use to filter results.

$sel:values:SearchDevicesFilter', searchDevicesFilter_values - The values to use to filter results.

searchDevicesFilter_name :: Lens' SearchDevicesFilter Text

The name to use to filter results.

searchDevicesFilter_values :: Lens' SearchDevicesFilter (NonEmpty Text)

The values to use to filter results.

SearchQuantumTasksFilter

data SearchQuantumTasksFilter

A filter to use to search for tasks.

See: newSearchQuantumTasksFilter smart constructor.

Constructors

SearchQuantumTasksFilter'

Fields

name :: Text

The name of the device used for the task.

operator :: SearchQuantumTasksFilterOperator

An operator to use in the filter.

values :: NonEmpty Text

The values to use for the filter.

newSearchQuantumTasksFilter

Arguments

:: Text
$sel:name:SearchQuantumTasksFilter'

-> SearchQuantumTasksFilterOperator
$sel:operator:SearchQuantumTasksFilter'

-> NonEmpty Text
$sel:values:SearchQuantumTasksFilter'

-> SearchQuantumTasksFilter

Create a value of SearchQuantumTasksFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:SearchQuantumTasksFilter', searchQuantumTasksFilter_name - The name of the device used for the task.

$sel:operator:SearchQuantumTasksFilter', searchQuantumTasksFilter_operator - An operator to use in the filter.

$sel:values:SearchQuantumTasksFilter', searchQuantumTasksFilter_values - The values to use for the filter.

searchQuantumTasksFilter_name :: Lens' SearchQuantumTasksFilter Text

The name of the device used for the task.

searchQuantumTasksFilter_operator :: Lens' SearchQuantumTasksFilter SearchQuantumTasksFilterOperator

An operator to use in the filter.

searchQuantumTasksFilter_values :: Lens' SearchQuantumTasksFilter (NonEmpty Text)

The values to use for the filter.