Network.AWS.Braket.Types.QuantumTaskSummary

Description

Documentation

data QuantumTaskSummary

Includes information about a quantum task.

See: newQuantumTaskSummary smart constructor.

Constructors

QuantumTaskSummary'

Fields

endedAt :: Maybe POSIX

The time at which the task finished.

tags :: Maybe (HashMap Text Text)

Displays the key, value pairs of tags associated with this quantum task.

createdAt :: POSIX

The time at which the task was created.

deviceArn :: Text

The ARN of the device the task ran on.

outputS3Bucket :: Text

The S3 bucket where the task result file is stored..

outputS3Directory :: Text

The folder in the S3 bucket where the task result file is stored.

quantumTaskArn :: Text

The ARN of the task.

shots :: Integer

The shots used for the task.

status :: QuantumTaskStatus

The status of the task.

newQuantumTaskSummary

Arguments

:: UTCTime
$sel:createdAt:QuantumTaskSummary'

-> Text
$sel:deviceArn:QuantumTaskSummary'

-> Text
$sel:outputS3Bucket:QuantumTaskSummary'

-> Text
$sel:outputS3Directory:QuantumTaskSummary'

-> Text
$sel:quantumTaskArn:QuantumTaskSummary'

-> Integer
$sel:shots:QuantumTaskSummary'

-> QuantumTaskStatus
$sel:status:QuantumTaskSummary'

-> QuantumTaskSummary

Create a value of QuantumTaskSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endedAt:QuantumTaskSummary', quantumTaskSummary_endedAt - The time at which the task finished.

$sel:tags:QuantumTaskSummary', quantumTaskSummary_tags - Displays the key, value pairs of tags associated with this quantum task.

$sel:createdAt:QuantumTaskSummary', quantumTaskSummary_createdAt - The time at which the task was created.

$sel:deviceArn:QuantumTaskSummary', quantumTaskSummary_deviceArn - The ARN of the device the task ran on.

$sel:outputS3Bucket:QuantumTaskSummary', quantumTaskSummary_outputS3Bucket - The S3 bucket where the task result file is stored..

$sel:outputS3Directory:QuantumTaskSummary', quantumTaskSummary_outputS3Directory - The folder in the S3 bucket where the task result file is stored.

$sel:quantumTaskArn:QuantumTaskSummary', quantumTaskSummary_quantumTaskArn - The ARN of the task.

$sel:shots:QuantumTaskSummary', quantumTaskSummary_shots - The shots used for the task.

$sel:status:QuantumTaskSummary', quantumTaskSummary_status - The status of the task.

quantumTaskSummary_endedAt :: Lens' QuantumTaskSummary (Maybe UTCTime)

The time at which the task finished.

quantumTaskSummary_tags :: Lens' QuantumTaskSummary (Maybe (HashMap Text Text))

Displays the key, value pairs of tags associated with this quantum task.

quantumTaskSummary_createdAt :: Lens' QuantumTaskSummary UTCTime

The time at which the task was created.

quantumTaskSummary_deviceArn :: Lens' QuantumTaskSummary Text

The ARN of the device the task ran on.

quantumTaskSummary_outputS3Bucket :: Lens' QuantumTaskSummary Text

The S3 bucket where the task result file is stored..

quantumTaskSummary_outputS3Directory :: Lens' QuantumTaskSummary Text

The folder in the S3 bucket where the task result file is stored.

quantumTaskSummary_quantumTaskArn :: Lens' QuantumTaskSummary Text

The ARN of the task.

quantumTaskSummary_shots :: Lens' QuantumTaskSummary Integer

The shots used for the task.

quantumTaskSummary_status :: Lens' QuantumTaskSummary QuantumTaskStatus

The status of the task.