Network.AWS.Budgets.CreateBudgetAction

Description

Creates a budget action.

Creating a Request

data CreateBudgetAction

See: newCreateBudgetAction smart constructor.

Constructors

CreateBudgetAction'

Fields

accountId :: Text

budgetName :: Text

notificationType :: NotificationType

actionType :: ActionType

The type of action. This defines the type of tasks that can be carried out by this action. This field also determines the format for definition.

actionThreshold :: ActionThreshold

definition :: Definition

executionRoleArn :: Text

The role passed for action execution and reversion. Roles and actions must be in the same account.

approvalModel :: ApprovalModel

This specifies if the action needs manual or automatic approval.

subscribers :: NonEmpty Subscriber

newCreateBudgetAction

Arguments

:: Text
$sel:accountId:CreateBudgetAction'

-> Text
$sel:budgetName:CreateBudgetAction'

-> NotificationType
$sel:notificationType:CreateBudgetAction'

-> ActionType
$sel:actionType:CreateBudgetAction'

-> ActionThreshold
$sel:actionThreshold:CreateBudgetAction'

-> Definition
$sel:definition:CreateBudgetAction'

-> Text
$sel:executionRoleArn:CreateBudgetAction'

-> ApprovalModel
$sel:approvalModel:CreateBudgetAction'

-> NonEmpty Subscriber
$sel:subscribers:CreateBudgetAction'

-> CreateBudgetAction

Create a value of CreateBudgetAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:CreateBudgetAction', createBudgetAction_accountId - Undocumented member.

$sel:budgetName:CreateBudgetAction', createBudgetAction_budgetName - Undocumented member.

$sel:notificationType:CreateBudgetAction', createBudgetAction_notificationType - Undocumented member.

$sel:actionType:CreateBudgetAction', createBudgetAction_actionType - The type of action. This defines the type of tasks that can be carried out by this action. This field also determines the format for definition.

$sel:actionThreshold:CreateBudgetAction', createBudgetAction_actionThreshold - Undocumented member.

$sel:definition:CreateBudgetAction', createBudgetAction_definition - Undocumented member.

$sel:executionRoleArn:CreateBudgetAction', createBudgetAction_executionRoleArn - The role passed for action execution and reversion. Roles and actions must be in the same account.

$sel:approvalModel:CreateBudgetAction', createBudgetAction_approvalModel - This specifies if the action needs manual or automatic approval.

$sel:subscribers:CreateBudgetAction', createBudgetAction_subscribers - Undocumented member.

Request Lenses

createBudgetAction_accountId :: Lens' CreateBudgetAction Text

Undocumented member.

createBudgetAction_budgetName :: Lens' CreateBudgetAction Text

Undocumented member.

createBudgetAction_notificationType :: Lens' CreateBudgetAction NotificationType

Undocumented member.

createBudgetAction_actionType :: Lens' CreateBudgetAction ActionType

The type of action. This defines the type of tasks that can be carried out by this action. This field also determines the format for definition.

createBudgetAction_actionThreshold :: Lens' CreateBudgetAction ActionThreshold

Undocumented member.

createBudgetAction_definition :: Lens' CreateBudgetAction Definition

Undocumented member.

createBudgetAction_executionRoleArn :: Lens' CreateBudgetAction Text

The role passed for action execution and reversion. Roles and actions must be in the same account.

createBudgetAction_approvalModel :: Lens' CreateBudgetAction ApprovalModel

This specifies if the action needs manual or automatic approval.

createBudgetAction_subscribers :: Lens' CreateBudgetAction (NonEmpty Subscriber)

Undocumented member.

Destructuring the Response

data CreateBudgetActionResponse

See: newCreateBudgetActionResponse smart constructor.

Constructors

CreateBudgetActionResponse'

Fields

httpStatus :: Int

The response's http status code.

accountId :: Text

budgetName :: Text

actionId :: Text

A system-generated universally unique identifier (UUID) for the action.

newCreateBudgetActionResponse

Arguments

:: Int
$sel:httpStatus:CreateBudgetActionResponse'

-> Text
$sel:accountId:CreateBudgetAction'

-> Text
$sel:budgetName:CreateBudgetAction'

-> Text
$sel:actionId:CreateBudgetActionResponse'

-> CreateBudgetActionResponse

Create a value of CreateBudgetActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateBudgetActionResponse', createBudgetActionResponse_httpStatus - The response's http status code.

$sel:accountId:CreateBudgetAction', createBudgetActionResponse_accountId - Undocumented member.

$sel:budgetName:CreateBudgetAction', createBudgetActionResponse_budgetName - Undocumented member.

$sel:actionId:CreateBudgetActionResponse', createBudgetActionResponse_actionId - A system-generated universally unique identifier (UUID) for the action.

Response Lenses

createBudgetActionResponse_httpStatus :: Lens' CreateBudgetActionResponse Int

The response's http status code.

createBudgetActionResponse_accountId :: Lens' CreateBudgetActionResponse Text

Undocumented member.

createBudgetActionResponse_budgetName :: Lens' CreateBudgetActionResponse Text

Undocumented member.

createBudgetActionResponse_actionId :: Lens' CreateBudgetActionResponse Text

A system-generated universally unique identifier (UUID) for the action.