Network.AWS.Budgets.DescribeBudget

Description

Describes a budget.

The Request Syntax section shows the BudgetLimit syntax. For PlannedBudgetLimits, see the Examples section.

Creating a Request

data DescribeBudget

Request of DescribeBudget

See: newDescribeBudget smart constructor.

Constructors

DescribeBudget'

Fields

accountId :: Text

The accountId that is associated with the budget that you want a description of.

budgetName :: Text

The name of the budget that you want a description of.

newDescribeBudget

Arguments

:: Text
$sel:accountId:DescribeBudget'

-> Text
$sel:budgetName:DescribeBudget'

-> DescribeBudget

Create a value of DescribeBudget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:DescribeBudget', describeBudget_accountId - The accountId that is associated with the budget that you want a description of.

$sel:budgetName:DescribeBudget', describeBudget_budgetName - The name of the budget that you want a description of.

Request Lenses

describeBudget_accountId :: Lens' DescribeBudget Text

The accountId that is associated with the budget that you want a description of.

describeBudget_budgetName :: Lens' DescribeBudget Text

The name of the budget that you want a description of.

Destructuring the Response

data DescribeBudgetResponse

Response of DescribeBudget

See: newDescribeBudgetResponse smart constructor.

Constructors

DescribeBudgetResponse'

Fields

budget :: Maybe Budget

The description of the budget.

httpStatus :: Int

The response's http status code.

newDescribeBudgetResponse

Arguments

:: Int
$sel:httpStatus:DescribeBudgetResponse'

-> DescribeBudgetResponse

Create a value of DescribeBudgetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:budget:DescribeBudgetResponse', describeBudgetResponse_budget - The description of the budget.

$sel:httpStatus:DescribeBudgetResponse', describeBudgetResponse_httpStatus - The response's http status code.

Response Lenses

describeBudgetResponse_budget :: Lens' DescribeBudgetResponse (Maybe Budget)

The description of the budget.

describeBudgetResponse_httpStatus :: Lens' DescribeBudgetResponse Int

The response's http status code.