Network.AWS.Budgets.DescribeBudgetAction

Description

Describes a budget action detail.

Creating a Request

data DescribeBudgetAction

See: newDescribeBudgetAction smart constructor.

Constructors

DescribeBudgetAction'

Fields

accountId :: Text

budgetName :: Text

actionId :: Text

A system-generated universally unique identifier (UUID) for the action.

newDescribeBudgetAction

Arguments

:: Text
$sel:accountId:DescribeBudgetAction'

-> Text
$sel:budgetName:DescribeBudgetAction'

-> Text
$sel:actionId:DescribeBudgetAction'

-> DescribeBudgetAction

Create a value of DescribeBudgetAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:DescribeBudgetAction', describeBudgetAction_accountId - Undocumented member.

$sel:budgetName:DescribeBudgetAction', describeBudgetAction_budgetName - Undocumented member.

$sel:actionId:DescribeBudgetAction', describeBudgetAction_actionId - A system-generated universally unique identifier (UUID) for the action.

Request Lenses

describeBudgetAction_accountId :: Lens' DescribeBudgetAction Text

Undocumented member.

describeBudgetAction_budgetName :: Lens' DescribeBudgetAction Text

Undocumented member.

describeBudgetAction_actionId :: Lens' DescribeBudgetAction Text

A system-generated universally unique identifier (UUID) for the action.

Destructuring the Response

data DescribeBudgetActionResponse

See: newDescribeBudgetActionResponse smart constructor.

Constructors

DescribeBudgetActionResponse'

Fields

httpStatus :: Int

The response's http status code.

accountId :: Text

budgetName :: Text

action :: Action

A budget action resource.

newDescribeBudgetActionResponse

Arguments

:: Int
$sel:httpStatus:DescribeBudgetActionResponse'

-> Text
$sel:accountId:DescribeBudgetAction'

-> Text
$sel:budgetName:DescribeBudgetAction'

-> Action
$sel:action:DescribeBudgetActionResponse'

-> DescribeBudgetActionResponse

Create a value of DescribeBudgetActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeBudgetActionResponse', describeBudgetActionResponse_httpStatus - The response's http status code.

$sel:accountId:DescribeBudgetAction', describeBudgetActionResponse_accountId - Undocumented member.

$sel:budgetName:DescribeBudgetAction', describeBudgetActionResponse_budgetName - Undocumented member.

$sel:action:DescribeBudgetActionResponse', describeBudgetActionResponse_action - A budget action resource.

Response Lenses

describeBudgetActionResponse_httpStatus :: Lens' DescribeBudgetActionResponse Int

The response's http status code.

describeBudgetActionResponse_accountId :: Lens' DescribeBudgetActionResponse Text

Undocumented member.

describeBudgetActionResponse_budgetName :: Lens' DescribeBudgetActionResponse Text

Undocumented member.

describeBudgetActionResponse_action :: Lens' DescribeBudgetActionResponse Action

A budget action resource.