Network.AWS.Budgets.DescribeBudgetActionHistories

Description

Describes a budget action history detail.

This operation returns paginated results.

Creating a Request

data DescribeBudgetActionHistories

See: newDescribeBudgetActionHistories smart constructor.

Constructors

DescribeBudgetActionHistories'

Fields

nextToken :: Maybe Text

maxResults :: Maybe Natural

timePeriod :: Maybe TimePeriod

accountId :: Text

budgetName :: Text

actionId :: Text

A system-generated universally unique identifier (UUID) for the action.

newDescribeBudgetActionHistories

Arguments

:: Text
$sel:accountId:DescribeBudgetActionHistories'

-> Text
$sel:budgetName:DescribeBudgetActionHistories'

-> Text
$sel:actionId:DescribeBudgetActionHistories'

-> DescribeBudgetActionHistories

Create a value of DescribeBudgetActionHistories with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeBudgetActionHistories', describeBudgetActionHistories_nextToken - Undocumented member.

$sel:maxResults:DescribeBudgetActionHistories', describeBudgetActionHistories_maxResults - Undocumented member.

$sel:timePeriod:DescribeBudgetActionHistories', describeBudgetActionHistories_timePeriod - Undocumented member.

$sel:accountId:DescribeBudgetActionHistories', describeBudgetActionHistories_accountId - Undocumented member.

$sel:budgetName:DescribeBudgetActionHistories', describeBudgetActionHistories_budgetName - Undocumented member.

$sel:actionId:DescribeBudgetActionHistories', describeBudgetActionHistories_actionId - A system-generated universally unique identifier (UUID) for the action.

Request Lenses

describeBudgetActionHistories_nextToken :: Lens' DescribeBudgetActionHistories (Maybe Text)

Undocumented member.

describeBudgetActionHistories_maxResults :: Lens' DescribeBudgetActionHistories (Maybe Natural)

Undocumented member.

describeBudgetActionHistories_timePeriod :: Lens' DescribeBudgetActionHistories (Maybe TimePeriod)

Undocumented member.

describeBudgetActionHistories_accountId :: Lens' DescribeBudgetActionHistories Text

Undocumented member.

describeBudgetActionHistories_budgetName :: Lens' DescribeBudgetActionHistories Text

Undocumented member.

describeBudgetActionHistories_actionId :: Lens' DescribeBudgetActionHistories Text

A system-generated universally unique identifier (UUID) for the action.

Destructuring the Response

data DescribeBudgetActionHistoriesResponse

See: newDescribeBudgetActionHistoriesResponse smart constructor.

Constructors

DescribeBudgetActionHistoriesResponse'

Fields

nextToken :: Maybe Text

httpStatus :: Int

The response's http status code.

actionHistories :: [ActionHistory]

The historical record of the budget action resource.

newDescribeBudgetActionHistoriesResponse

Arguments

:: Int
$sel:httpStatus:DescribeBudgetActionHistoriesResponse'

-> DescribeBudgetActionHistoriesResponse

Create a value of DescribeBudgetActionHistoriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeBudgetActionHistories', describeBudgetActionHistoriesResponse_nextToken - Undocumented member.

$sel:httpStatus:DescribeBudgetActionHistoriesResponse', describeBudgetActionHistoriesResponse_httpStatus - The response's http status code.

$sel:actionHistories:DescribeBudgetActionHistoriesResponse', describeBudgetActionHistoriesResponse_actionHistories - The historical record of the budget action resource.

Response Lenses

describeBudgetActionHistoriesResponse_nextToken :: Lens' DescribeBudgetActionHistoriesResponse (Maybe Text)

Undocumented member.

describeBudgetActionHistoriesResponse_httpStatus :: Lens' DescribeBudgetActionHistoriesResponse Int

The response's http status code.

describeBudgetActionHistoriesResponse_actionHistories :: Lens' DescribeBudgetActionHistoriesResponse [ActionHistory]

The historical record of the budget action resource.