Network.AWS.Budgets.Lens

Description

Operations

DescribeBudgetActionsForAccount

describeBudgetActionsForAccount_nextToken :: Lens' DescribeBudgetActionsForAccount (Maybe Text)

Undocumented member.

describeBudgetActionsForAccount_maxResults :: Lens' DescribeBudgetActionsForAccount (Maybe Natural)

Undocumented member.

describeBudgetActionsForAccount_accountId :: Lens' DescribeBudgetActionsForAccount Text

Undocumented member.

describeBudgetActionsForAccountResponse_nextToken :: Lens' DescribeBudgetActionsForAccountResponse (Maybe Text)

Undocumented member.

describeBudgetActionsForAccountResponse_httpStatus :: Lens' DescribeBudgetActionsForAccountResponse Int

The response's http status code.

describeBudgetActionsForAccountResponse_actions :: Lens' DescribeBudgetActionsForAccountResponse [Action]

A list of the budget action resources information.

CreateBudgetAction

createBudgetAction_accountId :: Lens' CreateBudgetAction Text

Undocumented member.

createBudgetAction_budgetName :: Lens' CreateBudgetAction Text

Undocumented member.

createBudgetAction_notificationType :: Lens' CreateBudgetAction NotificationType

Undocumented member.

createBudgetAction_actionType :: Lens' CreateBudgetAction ActionType

The type of action. This defines the type of tasks that can be carried out by this action. This field also determines the format for definition.

createBudgetAction_actionThreshold :: Lens' CreateBudgetAction ActionThreshold

Undocumented member.

createBudgetAction_definition :: Lens' CreateBudgetAction Definition

Undocumented member.

createBudgetAction_executionRoleArn :: Lens' CreateBudgetAction Text

The role passed for action execution and reversion. Roles and actions must be in the same account.

createBudgetAction_approvalModel :: Lens' CreateBudgetAction ApprovalModel

This specifies if the action needs manual or automatic approval.

createBudgetAction_subscribers :: Lens' CreateBudgetAction (NonEmpty Subscriber)

Undocumented member.

createBudgetActionResponse_httpStatus :: Lens' CreateBudgetActionResponse Int

The response's http status code.

createBudgetActionResponse_accountId :: Lens' CreateBudgetActionResponse Text

Undocumented member.

createBudgetActionResponse_budgetName :: Lens' CreateBudgetActionResponse Text

Undocumented member.

createBudgetActionResponse_actionId :: Lens' CreateBudgetActionResponse Text

A system-generated universally unique identifier (UUID) for the action.

ExecuteBudgetAction

executeBudgetAction_accountId :: Lens' ExecuteBudgetAction Text

Undocumented member.

executeBudgetAction_budgetName :: Lens' ExecuteBudgetAction Text

Undocumented member.

executeBudgetAction_actionId :: Lens' ExecuteBudgetAction Text

A system-generated universally unique identifier (UUID) for the action.

executeBudgetAction_executionType :: Lens' ExecuteBudgetAction ExecutionType

The type of execution.

executeBudgetActionResponse_httpStatus :: Lens' ExecuteBudgetActionResponse Int

The response's http status code.

executeBudgetActionResponse_accountId :: Lens' ExecuteBudgetActionResponse Text

Undocumented member.

executeBudgetActionResponse_budgetName :: Lens' ExecuteBudgetActionResponse Text

Undocumented member.

executeBudgetActionResponse_actionId :: Lens' ExecuteBudgetActionResponse Text

A system-generated universally unique identifier (UUID) for the action.

executeBudgetActionResponse_executionType :: Lens' ExecuteBudgetActionResponse ExecutionType

The type of execution.

DescribeBudgetAction

describeBudgetAction_accountId :: Lens' DescribeBudgetAction Text

Undocumented member.

describeBudgetAction_budgetName :: Lens' DescribeBudgetAction Text

Undocumented member.

describeBudgetAction_actionId :: Lens' DescribeBudgetAction Text

A system-generated universally unique identifier (UUID) for the action.

describeBudgetActionResponse_httpStatus :: Lens' DescribeBudgetActionResponse Int

The response's http status code.

describeBudgetActionResponse_accountId :: Lens' DescribeBudgetActionResponse Text

Undocumented member.

describeBudgetActionResponse_budgetName :: Lens' DescribeBudgetActionResponse Text

Undocumented member.

describeBudgetActionResponse_action :: Lens' DescribeBudgetActionResponse Action

A budget action resource.

UpdateBudgetAction

updateBudgetAction_subscribers :: Lens' UpdateBudgetAction (Maybe (NonEmpty Subscriber))

Undocumented member.

updateBudgetAction_executionRoleArn :: Lens' UpdateBudgetAction (Maybe Text)

The role passed for action execution and reversion. Roles and actions must be in the same account.

updateBudgetAction_approvalModel :: Lens' UpdateBudgetAction (Maybe ApprovalModel)

This specifies if the action needs manual or automatic approval.

updateBudgetAction_notificationType :: Lens' UpdateBudgetAction (Maybe NotificationType)

Undocumented member.

updateBudgetAction_actionThreshold :: Lens' UpdateBudgetAction (Maybe ActionThreshold)

Undocumented member.

updateBudgetAction_definition :: Lens' UpdateBudgetAction (Maybe Definition)

Undocumented member.

updateBudgetAction_accountId :: Lens' UpdateBudgetAction Text

Undocumented member.

updateBudgetAction_budgetName :: Lens' UpdateBudgetAction Text

Undocumented member.

updateBudgetAction_actionId :: Lens' UpdateBudgetAction Text

A system-generated universally unique identifier (UUID) for the action.

updateBudgetActionResponse_httpStatus :: Lens' UpdateBudgetActionResponse Int

The response's http status code.

updateBudgetActionResponse_accountId :: Lens' UpdateBudgetActionResponse Text

Undocumented member.

updateBudgetActionResponse_budgetName :: Lens' UpdateBudgetActionResponse Text

Undocumented member.

updateBudgetActionResponse_oldAction :: Lens' UpdateBudgetActionResponse Action

The previous action resource information.

updateBudgetActionResponse_newAction :: Lens' UpdateBudgetActionResponse Action

The updated action resource information.

DeleteBudgetAction

deleteBudgetAction_accountId :: Lens' DeleteBudgetAction Text

Undocumented member.

deleteBudgetAction_budgetName :: Lens' DeleteBudgetAction Text

Undocumented member.

deleteBudgetAction_actionId :: Lens' DeleteBudgetAction Text

A system-generated universally unique identifier (UUID) for the action.

deleteBudgetActionResponse_httpStatus :: Lens' DeleteBudgetActionResponse Int

The response's http status code.

deleteBudgetActionResponse_accountId :: Lens' DeleteBudgetActionResponse Text

Undocumented member.

deleteBudgetActionResponse_budgetName :: Lens' DeleteBudgetActionResponse Text

Undocumented member.

deleteBudgetActionResponse_action :: Lens' DeleteBudgetActionResponse Action

Undocumented member.

DescribeBudgetActionHistories

describeBudgetActionHistories_nextToken :: Lens' DescribeBudgetActionHistories (Maybe Text)

Undocumented member.

describeBudgetActionHistories_maxResults :: Lens' DescribeBudgetActionHistories (Maybe Natural)

Undocumented member.

describeBudgetActionHistories_timePeriod :: Lens' DescribeBudgetActionHistories (Maybe TimePeriod)

Undocumented member.

describeBudgetActionHistories_accountId :: Lens' DescribeBudgetActionHistories Text

Undocumented member.

describeBudgetActionHistories_budgetName :: Lens' DescribeBudgetActionHistories Text

Undocumented member.

describeBudgetActionHistories_actionId :: Lens' DescribeBudgetActionHistories Text

A system-generated universally unique identifier (UUID) for the action.

describeBudgetActionHistoriesResponse_nextToken :: Lens' DescribeBudgetActionHistoriesResponse (Maybe Text)

Undocumented member.

describeBudgetActionHistoriesResponse_httpStatus :: Lens' DescribeBudgetActionHistoriesResponse Int

The response's http status code.

describeBudgetActionHistoriesResponse_actionHistories :: Lens' DescribeBudgetActionHistoriesResponse [ActionHistory]

The historical record of the budget action resource.

DescribeSubscribersForNotification

describeSubscribersForNotification_nextToken :: Lens' DescribeSubscribersForNotification (Maybe Text)

The pagination token that you include in your request to indicate the next set of results that you want to retrieve.

describeSubscribersForNotification_maxResults :: Lens' DescribeSubscribersForNotification (Maybe Natural)

An optional integer that represents how many entries a paginated response contains. The maximum is 100.

describeSubscribersForNotification_accountId :: Lens' DescribeSubscribersForNotification Text

The accountId that is associated with the budget whose subscribers you want descriptions of.

describeSubscribersForNotification_budgetName :: Lens' DescribeSubscribersForNotification Text

The name of the budget whose subscribers you want descriptions of.

describeSubscribersForNotification_notification :: Lens' DescribeSubscribersForNotification Notification

The notification whose subscribers you want to list.

describeSubscribersForNotificationResponse_nextToken :: Lens' DescribeSubscribersForNotificationResponse (Maybe Text)

The pagination token in the service response that indicates the next set of results that you can retrieve.

describeSubscribersForNotificationResponse_subscribers :: Lens' DescribeSubscribersForNotificationResponse (Maybe (NonEmpty Subscriber))

A list of subscribers that are associated with a notification.

describeSubscribersForNotificationResponse_httpStatus :: Lens' DescribeSubscribersForNotificationResponse Int

The response's http status code.

UpdateBudget

updateBudget_accountId :: Lens' UpdateBudget Text

The accountId that is associated with the budget that you want to update.

updateBudget_newBudget :: Lens' UpdateBudget Budget

The budget that you want to update your budget to.

updateBudgetResponse_httpStatus :: Lens' UpdateBudgetResponse Int

The response's http status code.

DeleteNotification

deleteNotification_accountId :: Lens' DeleteNotification Text

The accountId that is associated with the budget whose notification you want to delete.

deleteNotification_budgetName :: Lens' DeleteNotification Text

The name of the budget whose notification you want to delete.

deleteNotification_notification :: Lens' DeleteNotification Notification

The notification that you want to delete.

deleteNotificationResponse_httpStatus :: Lens' DeleteNotificationResponse Int

The response's http status code.

DeleteBudget

deleteBudget_accountId :: Lens' DeleteBudget Text

The accountId that is associated with the budget that you want to delete.

deleteBudget_budgetName :: Lens' DeleteBudget Text

The name of the budget that you want to delete.

deleteBudgetResponse_httpStatus :: Lens' DeleteBudgetResponse Int

The response's http status code.

UpdateNotification

updateNotification_accountId :: Lens' UpdateNotification Text

The accountId that is associated with the budget whose notification you want to update.

updateNotification_budgetName :: Lens' UpdateNotification Text

The name of the budget whose notification you want to update.

updateNotification_oldNotification :: Lens' UpdateNotification Notification

The previous notification that is associated with a budget.

updateNotification_newNotification :: Lens' UpdateNotification Notification

The updated notification to be associated with a budget.

updateNotificationResponse_httpStatus :: Lens' UpdateNotificationResponse Int

The response's http status code.

CreateBudget

createBudget_notificationsWithSubscribers :: Lens' CreateBudget (Maybe [NotificationWithSubscribers])

A notification that you want to associate with a budget. A budget can have up to five notifications, and each notification can have one SNS subscriber and up to 10 email subscribers. If you include notifications and subscribers in your CreateBudget call, AWS creates the notifications and subscribers for you.

createBudget_accountId :: Lens' CreateBudget Text

The accountId that is associated with the budget.

createBudget_budget :: Lens' CreateBudget Budget

The budget object that you want to create.

createBudgetResponse_httpStatus :: Lens' CreateBudgetResponse Int

The response's http status code.

CreateNotification

createNotification_accountId :: Lens' CreateNotification Text

The accountId that is associated with the budget that you want to create a notification for.

createNotification_budgetName :: Lens' CreateNotification Text

The name of the budget that you want AWS to notify you about. Budget names must be unique within an account.

createNotification_notification :: Lens' CreateNotification Notification

The notification that you want to create.

createNotification_subscribers :: Lens' CreateNotification (NonEmpty Subscriber)

A list of subscribers that you want to associate with the notification. Each notification can have one SNS subscriber and up to 10 email subscribers.

createNotificationResponse_httpStatus :: Lens' CreateNotificationResponse Int

The response's http status code.

CreateSubscriber

createSubscriber_accountId :: Lens' CreateSubscriber Text

The accountId that is associated with the budget that you want to create a subscriber for.

createSubscriber_budgetName :: Lens' CreateSubscriber Text

The name of the budget that you want to subscribe to. Budget names must be unique within an account.

createSubscriber_notification :: Lens' CreateSubscriber Notification

The notification that you want to create a subscriber for.

createSubscriber_subscriber :: Lens' CreateSubscriber Subscriber

The subscriber that you want to associate with a budget notification.

createSubscriberResponse_httpStatus :: Lens' CreateSubscriberResponse Int

The response's http status code.

DescribeBudgets

describeBudgets_nextToken :: Lens' DescribeBudgets (Maybe Text)

The pagination token that you include in your request to indicate the next set of results that you want to retrieve.

describeBudgets_maxResults :: Lens' DescribeBudgets (Maybe Natural)

An optional integer that represents how many entries a paginated response contains. The maximum is 100.

describeBudgets_accountId :: Lens' DescribeBudgets Text

The accountId that is associated with the budgets that you want descriptions of.

describeBudgetsResponse_nextToken :: Lens' DescribeBudgetsResponse (Maybe Text)

The pagination token in the service response that indicates the next set of results that you can retrieve.

describeBudgetsResponse_budgets :: Lens' DescribeBudgetsResponse (Maybe [Budget])

A list of budgets.

describeBudgetsResponse_httpStatus :: Lens' DescribeBudgetsResponse Int

The response's http status code.

DeleteSubscriber

deleteSubscriber_accountId :: Lens' DeleteSubscriber Text

The accountId that is associated with the budget whose subscriber you want to delete.

deleteSubscriber_budgetName :: Lens' DeleteSubscriber Text

The name of the budget whose subscriber you want to delete.

deleteSubscriber_notification :: Lens' DeleteSubscriber Notification

The notification whose subscriber you want to delete.

deleteSubscriber_subscriber :: Lens' DeleteSubscriber Subscriber

The subscriber that you want to delete.

deleteSubscriberResponse_httpStatus :: Lens' DeleteSubscriberResponse Int

The response's http status code.

UpdateSubscriber

updateSubscriber_accountId :: Lens' UpdateSubscriber Text

The accountId that is associated with the budget whose subscriber you want to update.

updateSubscriber_budgetName :: Lens' UpdateSubscriber Text

The name of the budget whose subscriber you want to update.

updateSubscriber_notification :: Lens' UpdateSubscriber Notification

The notification whose subscriber you want to update.

updateSubscriber_oldSubscriber :: Lens' UpdateSubscriber Subscriber

The previous subscriber that is associated with a budget notification.

updateSubscriber_newSubscriber :: Lens' UpdateSubscriber Subscriber

The updated subscriber that is associated with a budget notification.

updateSubscriberResponse_httpStatus :: Lens' UpdateSubscriberResponse Int

The response's http status code.

DescribeBudgetActionsForBudget

describeBudgetActionsForBudget_nextToken :: Lens' DescribeBudgetActionsForBudget (Maybe Text)

Undocumented member.

describeBudgetActionsForBudget_maxResults :: Lens' DescribeBudgetActionsForBudget (Maybe Natural)

Undocumented member.

describeBudgetActionsForBudget_accountId :: Lens' DescribeBudgetActionsForBudget Text

Undocumented member.

describeBudgetActionsForBudget_budgetName :: Lens' DescribeBudgetActionsForBudget Text

Undocumented member.

describeBudgetActionsForBudgetResponse_nextToken :: Lens' DescribeBudgetActionsForBudgetResponse (Maybe Text)

Undocumented member.

describeBudgetActionsForBudgetResponse_httpStatus :: Lens' DescribeBudgetActionsForBudgetResponse Int

The response's http status code.

describeBudgetActionsForBudgetResponse_actions :: Lens' DescribeBudgetActionsForBudgetResponse [Action]

A list of the budget action resources information.

DescribeBudget

describeBudget_accountId :: Lens' DescribeBudget Text

The accountId that is associated with the budget that you want a description of.

describeBudget_budgetName :: Lens' DescribeBudget Text

The name of the budget that you want a description of.

describeBudgetResponse_budget :: Lens' DescribeBudgetResponse (Maybe Budget)

The description of the budget.

describeBudgetResponse_httpStatus :: Lens' DescribeBudgetResponse Int

The response's http status code.

DescribeNotificationsForBudget

describeNotificationsForBudget_nextToken :: Lens' DescribeNotificationsForBudget (Maybe Text)

The pagination token that you include in your request to indicate the next set of results that you want to retrieve.

describeNotificationsForBudget_maxResults :: Lens' DescribeNotificationsForBudget (Maybe Natural)

An optional integer that represents how many entries a paginated response contains. The maximum is 100.

describeNotificationsForBudget_accountId :: Lens' DescribeNotificationsForBudget Text

The accountId that is associated with the budget whose notifications you want descriptions of.

describeNotificationsForBudget_budgetName :: Lens' DescribeNotificationsForBudget Text

The name of the budget whose notifications you want descriptions of.

describeNotificationsForBudgetResponse_nextToken :: Lens' DescribeNotificationsForBudgetResponse (Maybe Text)

The pagination token in the service response that indicates the next set of results that you can retrieve.

describeNotificationsForBudgetResponse_notifications :: Lens' DescribeNotificationsForBudgetResponse (Maybe [Notification])

A list of notifications that are associated with a budget.

describeNotificationsForBudgetResponse_httpStatus :: Lens' DescribeNotificationsForBudgetResponse Int

The response's http status code.

DescribeBudgetPerformanceHistory

describeBudgetPerformanceHistory_nextToken :: Lens' DescribeBudgetPerformanceHistory (Maybe Text)

Undocumented member.

describeBudgetPerformanceHistory_maxResults :: Lens' DescribeBudgetPerformanceHistory (Maybe Natural)

Undocumented member.

describeBudgetPerformanceHistory_timePeriod :: Lens' DescribeBudgetPerformanceHistory (Maybe TimePeriod)

Retrieves how often the budget went into an ALARM state for the specified time period.

describeBudgetPerformanceHistory_accountId :: Lens' DescribeBudgetPerformanceHistory Text

Undocumented member.

describeBudgetPerformanceHistory_budgetName :: Lens' DescribeBudgetPerformanceHistory Text

Undocumented member.

describeBudgetPerformanceHistoryResponse_budgetPerformanceHistory :: Lens' DescribeBudgetPerformanceHistoryResponse (Maybe BudgetPerformanceHistory)

The history of how often the budget has gone into an ALARM state.

For DAILY budgets, the history saves the state of the budget for the last 60 days. For MONTHLY budgets, the history saves the state of the budget for the current month plus the last 12 months. For QUARTERLY budgets, the history saves the state of the budget for the last four quarters.

describeBudgetPerformanceHistoryResponse_nextToken :: Lens' DescribeBudgetPerformanceHistoryResponse (Maybe Text)

Undocumented member.

describeBudgetPerformanceHistoryResponse_httpStatus :: Lens' DescribeBudgetPerformanceHistoryResponse Int

The response's http status code.

Types

Action

action_actionId :: Lens' Action Text

A system-generated universally unique identifier (UUID) for the action.

action_budgetName :: Lens' Action Text

Undocumented member.

action_notificationType :: Lens' Action NotificationType

Undocumented member.

action_actionType :: Lens' Action ActionType

The type of action. This defines the type of tasks that can be carried out by this action. This field also determines the format for definition.

action_actionThreshold :: Lens' Action ActionThreshold

The trigger threshold of the action.

action_definition :: Lens' Action Definition

Where you specify all of the type-specific parameters.

action_executionRoleArn :: Lens' Action Text

The role passed for action execution and reversion. Roles and actions must be in the same account.

action_approvalModel :: Lens' Action ApprovalModel

This specifies if the action needs manual or automatic approval.

action_status :: Lens' Action ActionStatus

The status of action.

action_subscribers :: Lens' Action (NonEmpty Subscriber)

Undocumented member.

ActionHistory

actionHistory_timestamp :: Lens' ActionHistory UTCTime

Undocumented member.

actionHistory_status :: Lens' ActionHistory ActionStatus

The status of action at the time of the event.

actionHistory_eventType :: Lens' ActionHistory EventType

This distinguishes between whether the events are triggered by the user or generated by the system.

actionHistory_actionHistoryDetails :: Lens' ActionHistory ActionHistoryDetails

The description of details of the event.

ActionHistoryDetails

actionHistoryDetails_message :: Lens' ActionHistoryDetails Text

Undocumented member.

actionHistoryDetails_action :: Lens' ActionHistoryDetails Action

The budget action resource.

ActionThreshold

actionThreshold_actionThresholdValue :: Lens' ActionThreshold Double

Undocumented member.

actionThreshold_actionThresholdType :: Lens' ActionThreshold ThresholdType

Undocumented member.

Budget

budget_timePeriod :: Lens' Budget (Maybe TimePeriod)

The period of time that is covered by a budget. The period has a start date and an end date. The start date must come before the end date. The end date must come before 06/15/87 00:00 UTC.

If you create your budget and don't specify a start date, AWS defaults to the start of your chosen time period (DAILY, MONTHLY, QUARTERLY, or ANNUALLY). For example, if you created your budget on January 24, 2018, chose DAILY, and didn't set a start date, AWS set your start date to 01/24/18 00:00 UTC. If you chose MONTHLY, AWS set your start date to 01/01/18 00:00 UTC. If you didn't specify an end date, AWS set your end date to 06/15/87 00:00 UTC. The defaults are the same for the AWS Billing and Cost Management console and the API.

You can change either date with the UpdateBudget operation.

After the end date, AWS deletes the budget and all associated notifications and subscribers.

budget_costFilters :: Lens' Budget (Maybe (HashMap Text [Text]))

The cost filters, such as service or tag, that are applied to a budget.

AWS Budgets supports the following services as a filter for RI budgets:

Amazon Elastic Compute Cloud - Compute

Amazon Redshift

Amazon Relational Database Service

Amazon ElastiCache

Amazon Elasticsearch Service

budget_costTypes :: Lens' Budget (Maybe CostTypes)

The types of costs that are included in this COST budget.

USAGE, RI_UTILIZATION, RI_COVERAGE, SAVINGS_PLANS_UTILIZATION, and SAVINGS_PLANS_COVERAGE budgets do not have CostTypes.

budget_plannedBudgetLimits :: Lens' Budget (Maybe (HashMap Text Spend))

A map containing multiple BudgetLimit, including current or future limits.

PlannedBudgetLimits is available for cost or usage budget and supports monthly and quarterly TimeUnit.

For monthly budgets, provide 12 months of PlannedBudgetLimits values. This must start from the current month and include the next 11 months. The key is the start of the month, UTC in epoch seconds.

For quarterly budgets, provide 4 quarters of PlannedBudgetLimits value entries in standard calendar quarter increments. This must start from the current quarter and include the next 3 quarters. The key is the start of the quarter, UTC in epoch seconds.

If the planned budget expires before 12 months for monthly or 4 quarters for quarterly, provide the PlannedBudgetLimits values only for the remaining periods.

If the budget begins at a date in the future, provide PlannedBudgetLimits values from the start date of the budget.

After all of the BudgetLimit values in PlannedBudgetLimits are used, the budget continues to use the last limit as the BudgetLimit. At that point, the planned budget provides the same experience as a fixed budget.

DescribeBudget and DescribeBudgets response along with PlannedBudgetLimits will also contain BudgetLimit representing the current month or quarter limit present in PlannedBudgetLimits. This only applies to budgets created with PlannedBudgetLimits. Budgets created without PlannedBudgetLimits will only contain BudgetLimit, and no PlannedBudgetLimits.

budget_calculatedSpend :: Lens' Budget (Maybe CalculatedSpend)

The actual and forecasted cost or usage that the budget tracks.

budget_budgetLimit :: Lens' Budget (Maybe Spend)

The total amount of cost, usage, RI utilization, RI coverage, Savings Plans utilization, or Savings Plans coverage that you want to track with your budget.

BudgetLimit is required for cost or usage budgets, but optional for RI or Savings Plans utilization or coverage budgets. RI and Savings Plans utilization or coverage budgets default to 100, which is the only valid value for RI or Savings Plans utilization or coverage budgets. You can't use BudgetLimit with PlannedBudgetLimits for CreateBudget and UpdateBudget actions.

budget_lastUpdatedTime :: Lens' Budget (Maybe UTCTime)

The last time that you updated this budget.

budget_budgetName :: Lens' Budget Text

The name of a budget. The name must be unique within an account. The : and \ characters aren't allowed in BudgetName.

budget_timeUnit :: Lens' Budget TimeUnit

The length of time until a budget resets the actual and forecasted spend.

budget_budgetType :: Lens' Budget BudgetType

Whether this budget tracks costs, usage, RI utilization, RI coverage, Savings Plans utilization, or Savings Plans coverage.

BudgetPerformanceHistory

budgetPerformanceHistory_budgetedAndActualAmountsList :: Lens' BudgetPerformanceHistory (Maybe [BudgetedAndActualAmounts])

A list of amounts of cost or usage that you created budgets for, compared to your actual costs or usage.

budgetPerformanceHistory_timeUnit :: Lens' BudgetPerformanceHistory (Maybe TimeUnit)

Undocumented member.

budgetPerformanceHistory_costFilters :: Lens' BudgetPerformanceHistory (Maybe (HashMap Text [Text]))

The history of the cost filters for a budget during the specified time period.

budgetPerformanceHistory_costTypes :: Lens' BudgetPerformanceHistory (Maybe CostTypes)

The history of the cost types for a budget during the specified time period.

budgetPerformanceHistory_budgetType :: Lens' BudgetPerformanceHistory (Maybe BudgetType)

Undocumented member.

budgetPerformanceHistory_budgetName :: Lens' BudgetPerformanceHistory (Maybe Text)

Undocumented member.

BudgetedAndActualAmounts

budgetedAndActualAmounts_timePeriod :: Lens' BudgetedAndActualAmounts (Maybe TimePeriod)

The time period covered by this budget comparison.

budgetedAndActualAmounts_budgetedAmount :: Lens' BudgetedAndActualAmounts (Maybe Spend)

The amount of cost or usage that you created the budget for.

budgetedAndActualAmounts_actualAmount :: Lens' BudgetedAndActualAmounts (Maybe Spend)

Your actual costs or usage for a budget period.

CalculatedSpend

calculatedSpend_forecastedSpend :: Lens' CalculatedSpend (Maybe Spend)

The amount of cost, usage, RI units, or Savings Plans units that you are forecasted to use.

calculatedSpend_actualSpend :: Lens' CalculatedSpend Spend

The amount of cost, usage, RI units, or Savings Plans units that you have used.

CostTypes

costTypes_includeSubscription :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes subscriptions.

The default value is true.

costTypes_useAmortized :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget uses the amortized rate.

The default value is false.

costTypes_includeCredit :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes credits.

The default value is true.

costTypes_useBlended :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget uses a blended rate.

The default value is false.

costTypes_includeSupport :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes support subscription fees.

The default value is true.

costTypes_includeRefund :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes refunds.

The default value is true.

costTypes_includeTax :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes taxes.

The default value is true.

costTypes_includeDiscount :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes discounts.

The default value is true.

costTypes_includeOtherSubscription :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes non-RI subscription costs.

The default value is true.

costTypes_includeUpfront :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes upfront RI costs.

The default value is true.

costTypes_includeRecurring :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes recurring fees such as monthly RI fees.

The default value is true.

Definition

definition_iamActionDefinition :: Lens' Definition (Maybe IamActionDefinition)

The AWS Identity and Access Management (IAM) action definition details.

definition_ssmActionDefinition :: Lens' Definition (Maybe SsmActionDefinition)

The AWS Systems Manager (SSM) action definition details.

definition_scpActionDefinition :: Lens' Definition (Maybe ScpActionDefinition)

The service control policies (SCPs) action definition details.

IamActionDefinition

iamActionDefinition_groups :: Lens' IamActionDefinition (Maybe (NonEmpty Text))

A list of groups to be attached. There must be at least one group.

iamActionDefinition_roles :: Lens' IamActionDefinition (Maybe (NonEmpty Text))

A list of roles to be attached. There must be at least one role.

iamActionDefinition_users :: Lens' IamActionDefinition (Maybe (NonEmpty Text))

A list of users to be attached. There must be at least one user.

iamActionDefinition_policyArn :: Lens' IamActionDefinition Text

The Amazon Resource Name (ARN) of the policy to be attached.

Notification

notification_notificationState :: Lens' Notification (Maybe NotificationState)

Whether this notification is in alarm. If a budget notification is in the ALARM state, you have passed the set threshold for the budget.

notification_thresholdType :: Lens' Notification (Maybe ThresholdType)

The type of threshold for a notification. For ABSOLUTE_VALUE thresholds, AWS notifies you when you go over or are forecasted to go over your total cost threshold. For PERCENTAGE thresholds, AWS notifies you when you go over or are forecasted to go over a certain percentage of your forecasted spend. For example, if you have a budget for 200 dollars and you have a PERCENTAGE threshold of 80%, AWS notifies you when you go over 160 dollars.

notification_notificationType :: Lens' Notification NotificationType

Whether the notification is for how much you have spent (ACTUAL) or for how much you're forecasted to spend (FORECASTED).

notification_comparisonOperator :: Lens' Notification ComparisonOperator

The comparison that is used for this notification.

notification_threshold :: Lens' Notification Double

The threshold that is associated with a notification. Thresholds are always a percentage, and many customers find value being alerted between 50% - 200% of the budgeted amount. The maximum limit for your threshold is 1,000,000% above the budgeted amount.

NotificationWithSubscribers

notificationWithSubscribers_notification :: Lens' NotificationWithSubscribers Notification

The notification that is associated with a budget.

notificationWithSubscribers_subscribers :: Lens' NotificationWithSubscribers (NonEmpty Subscriber)

A list of subscribers who are subscribed to this notification.

ScpActionDefinition

scpActionDefinition_policyId :: Lens' ScpActionDefinition Text

The policy ID attached.

scpActionDefinition_targetIds :: Lens' ScpActionDefinition (NonEmpty Text)

A list of target IDs.

Spend

spend_amount :: Lens' Spend Text

The cost or usage amount that is associated with a budget forecast, actual spend, or budget threshold.

spend_unit :: Lens' Spend Text

The unit of measurement that is used for the budget forecast, actual spend, or budget threshold, such as dollars or GB.

SsmActionDefinition

ssmActionDefinition_actionSubType :: Lens' SsmActionDefinition ActionSubType

The action subType.

ssmActionDefinition_region :: Lens' SsmActionDefinition Text

The Region to run the SSM document.

ssmActionDefinition_instanceIds :: Lens' SsmActionDefinition (NonEmpty Text)

The EC2 and RDS instance IDs.

Subscriber

subscriber_subscriptionType :: Lens' Subscriber SubscriptionType

The type of notification that AWS sends to a subscriber.

subscriber_address :: Lens' Subscriber Text

The address that AWS sends budget notifications to, either an SNS topic or an email.

When you create a subscriber, the value of Address can't contain line breaks.

TimePeriod

timePeriod_end :: Lens' TimePeriod (Maybe UTCTime)

The end date for a budget. If you didn't specify an end date, AWS set your end date to 06/15/87 00:00 UTC. The defaults are the same for the AWS Billing and Cost Management console and the API.

After the end date, AWS deletes the budget and all associated notifications and subscribers. You can change your end date with the UpdateBudget operation.

timePeriod_start :: Lens' TimePeriod (Maybe UTCTime)

The start date for a budget. If you created your budget and didn't specify a start date, AWS defaults to the start of your chosen time period (DAILY, MONTHLY, QUARTERLY, or ANNUALLY). For example, if you created your budget on January 24, 2018, chose DAILY, and didn't set a start date, AWS set your start date to 01/24/18 00:00 UTC. If you chose MONTHLY, AWS set your start date to 01/01/18 00:00 UTC. The defaults are the same for the AWS Billing and Cost Management console and the API.

You can change your start date with the UpdateBudget operation.