Network.AWS.Budgets.Types

Description

Service Configuration

defaultService :: Service

API version 2016-10-20 of the Amazon Budgets SDK configuration.

Errors

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You are not authorized to use this operation with the given parameters.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

An error on the client occurred. Typically, the cause is an invalid input value.

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An error on the server occurred during the processing of your request. Try again later.

_ExpiredNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The pagination token expired.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

We can’t locate the resource that you specified.

_ResourceLockedException :: AsError a => Getting (First ServiceError) a ServiceError

The request was received and recognized by the server, but the server rejected that particular method for the requested resource.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The pagination token is invalid.

_DuplicateRecordException :: AsError a => Getting (First ServiceError) a ServiceError

The budget name already exists. Budget names must be unique within an account.

_CreationLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You've exceeded the notification or subscriber limit.

ActionStatus

newtype ActionStatus

Constructors

ActionStatus'

Fields

fromActionStatus :: Text

Bundled Patterns

pattern ActionStatus_EXECUTION_FAILURE :: ActionStatus

pattern ActionStatus_EXECUTION_IN_PROGRESS :: ActionStatus

pattern ActionStatus_EXECUTION_SUCCESS :: ActionStatus

pattern ActionStatus_PENDING :: ActionStatus

pattern ActionStatus_RESET_FAILURE :: ActionStatus

pattern ActionStatus_RESET_IN_PROGRESS :: ActionStatus

pattern ActionStatus_REVERSE_FAILURE :: ActionStatus

pattern ActionStatus_REVERSE_IN_PROGRESS :: ActionStatus

pattern ActionStatus_REVERSE_SUCCESS :: ActionStatus

pattern ActionStatus_STANDBY :: ActionStatus

ActionSubType

newtype ActionSubType

Constructors

ActionSubType'

Fields

fromActionSubType :: Text

Bundled Patterns

pattern ActionSubType_STOP_EC2_INSTANCES :: ActionSubType

pattern ActionSubType_STOP_RDS_INSTANCES :: ActionSubType

ActionType

newtype ActionType

Constructors

ActionType'

Fields

fromActionType :: Text

Bundled Patterns

pattern ActionType_APPLY_IAM_POLICY :: ActionType

pattern ActionType_APPLY_SCP_POLICY :: ActionType

pattern ActionType_RUN_SSM_DOCUMENTS :: ActionType

ApprovalModel

newtype ApprovalModel

Constructors

ApprovalModel'

Fields

fromApprovalModel :: Text

Bundled Patterns

pattern ApprovalModel_AUTOMATIC :: ApprovalModel

pattern ApprovalModel_MANUAL :: ApprovalModel

BudgetType

newtype BudgetType

The type of a budget. It must be one of the following types:

COST, USAGE, RI_UTILIZATION, RI_COVERAGE, SAVINGS_PLANS_UTILIZATION, or SAVINGS_PLANS_COVERAGE.

Constructors

BudgetType'

Fields

fromBudgetType :: Text

Bundled Patterns

pattern BudgetType_COST :: BudgetType

pattern BudgetType_RI_COVERAGE :: BudgetType

pattern BudgetType_RI_UTILIZATION :: BudgetType

pattern BudgetType_SAVINGS_PLANS_COVERAGE :: BudgetType

pattern BudgetType_SAVINGS_PLANS_UTILIZATION :: BudgetType

pattern BudgetType_USAGE :: BudgetType

ComparisonOperator

newtype ComparisonOperator

The comparison operator of a notification. Currently the service supports the following operators:

GREATER_THAN, LESS_THAN, EQUAL_TO

Constructors

ComparisonOperator'

Fields

fromComparisonOperator :: Text

Bundled Patterns

pattern ComparisonOperator_EQUAL_TO :: ComparisonOperator

pattern ComparisonOperator_GREATER_THAN :: ComparisonOperator

pattern ComparisonOperator_LESS_THAN :: ComparisonOperator

EventType

newtype EventType

Constructors

EventType'

Fields

fromEventType :: Text

Bundled Patterns

pattern EventType_CREATE_ACTION :: EventType

pattern EventType_DELETE_ACTION :: EventType

pattern EventType_EXECUTE_ACTION :: EventType

pattern EventType_SYSTEM :: EventType

pattern EventType_UPDATE_ACTION :: EventType

ExecutionType

newtype ExecutionType

Constructors

ExecutionType'

Fields

fromExecutionType :: Text

Bundled Patterns

pattern ExecutionType_APPROVE_BUDGET_ACTION :: ExecutionType

pattern ExecutionType_RESET_BUDGET_ACTION :: ExecutionType

pattern ExecutionType_RETRY_BUDGET_ACTION :: ExecutionType

pattern ExecutionType_REVERSE_BUDGET_ACTION :: ExecutionType

NotificationState

newtype NotificationState

Constructors

NotificationState'

Fields

fromNotificationState :: Text

Bundled Patterns

pattern NotificationState_ALARM :: NotificationState

pattern NotificationState_OK :: NotificationState

NotificationType

newtype NotificationType

The type of a notification. It must be ACTUAL or FORECASTED.

Constructors

NotificationType'

Fields

fromNotificationType :: Text

Bundled Patterns

pattern NotificationType_ACTUAL :: NotificationType

pattern NotificationType_FORECASTED :: NotificationType

SubscriptionType

newtype SubscriptionType

The subscription type of the subscriber. It can be SMS or EMAIL.

Constructors

SubscriptionType'

Fields

fromSubscriptionType :: Text

Bundled Patterns

pattern SubscriptionType_EMAIL :: SubscriptionType

pattern SubscriptionType_SNS :: SubscriptionType

ThresholdType

newtype ThresholdType

The type of threshold for a notification.

Constructors

ThresholdType'

Fields

fromThresholdType :: Text

Bundled Patterns

pattern ThresholdType_ABSOLUTE_VALUE :: ThresholdType

pattern ThresholdType_PERCENTAGE :: ThresholdType

TimeUnit

newtype TimeUnit

The time unit of the budget, such as MONTHLY or QUARTERLY.

Constructors

TimeUnit'

Fields

fromTimeUnit :: Text

Bundled Patterns

pattern TimeUnit_ANNUALLY :: TimeUnit

pattern TimeUnit_DAILY :: TimeUnit

pattern TimeUnit_MONTHLY :: TimeUnit

pattern TimeUnit_QUARTERLY :: TimeUnit

Action

data Action

A budget action resource.

See: newAction smart constructor.

Constructors

Action'

Fields

actionId :: Text

A system-generated universally unique identifier (UUID) for the action.

budgetName :: Text

notificationType :: NotificationType

actionType :: ActionType

The type of action. This defines the type of tasks that can be carried out by this action. This field also determines the format for definition.

actionThreshold :: ActionThreshold

The trigger threshold of the action.

definition :: Definition

Where you specify all of the type-specific parameters.

executionRoleArn :: Text

The role passed for action execution and reversion. Roles and actions must be in the same account.

approvalModel :: ApprovalModel

This specifies if the action needs manual or automatic approval.

status :: ActionStatus

The status of action.

subscribers :: NonEmpty Subscriber

newAction

Arguments

:: Text
$sel:actionId:Action'

-> Text
$sel:budgetName:Action'

-> NotificationType
$sel:notificationType:Action'

-> ActionType
$sel:actionType:Action'

-> ActionThreshold
$sel:actionThreshold:Action'

-> Definition
$sel:definition:Action'

-> Text
$sel:executionRoleArn:Action'

-> ApprovalModel
$sel:approvalModel:Action'

-> ActionStatus
$sel:status:Action'

-> NonEmpty Subscriber
$sel:subscribers:Action'

-> Action

Create a value of Action with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionId:Action', action_actionId - A system-generated universally unique identifier (UUID) for the action.

$sel:budgetName:Action', action_budgetName - Undocumented member.

$sel:notificationType:Action', action_notificationType - Undocumented member.

$sel:actionType:Action', action_actionType - The type of action. This defines the type of tasks that can be carried out by this action. This field also determines the format for definition.

$sel:actionThreshold:Action', action_actionThreshold - The trigger threshold of the action.

$sel:definition:Action', action_definition - Where you specify all of the type-specific parameters.

$sel:executionRoleArn:Action', action_executionRoleArn - The role passed for action execution and reversion. Roles and actions must be in the same account.

$sel:approvalModel:Action', action_approvalModel - This specifies if the action needs manual or automatic approval.

$sel:status:Action', action_status - The status of action.

$sel:subscribers:Action', action_subscribers - Undocumented member.

action_actionId :: Lens' Action Text

A system-generated universally unique identifier (UUID) for the action.

action_budgetName :: Lens' Action Text

Undocumented member.

action_notificationType :: Lens' Action NotificationType

Undocumented member.

action_actionType :: Lens' Action ActionType

The type of action. This defines the type of tasks that can be carried out by this action. This field also determines the format for definition.

action_actionThreshold :: Lens' Action ActionThreshold

The trigger threshold of the action.

action_definition :: Lens' Action Definition

Where you specify all of the type-specific parameters.

action_executionRoleArn :: Lens' Action Text

The role passed for action execution and reversion. Roles and actions must be in the same account.

action_approvalModel :: Lens' Action ApprovalModel

This specifies if the action needs manual or automatic approval.

action_status :: Lens' Action ActionStatus

The status of action.

action_subscribers :: Lens' Action (NonEmpty Subscriber)

Undocumented member.

ActionHistory

data ActionHistory

The historical records for a budget action.

See: newActionHistory smart constructor.

Constructors

ActionHistory'

Fields

timestamp :: POSIX

status :: ActionStatus

The status of action at the time of the event.

eventType :: EventType

This distinguishes between whether the events are triggered by the user or generated by the system.

actionHistoryDetails :: ActionHistoryDetails

The description of details of the event.

newActionHistory

Arguments

:: UTCTime
$sel:timestamp:ActionHistory'

-> ActionStatus
$sel:status:ActionHistory'

-> EventType
$sel:eventType:ActionHistory'

-> ActionHistoryDetails
$sel:actionHistoryDetails:ActionHistory'

-> ActionHistory

Create a value of ActionHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timestamp:ActionHistory', actionHistory_timestamp - Undocumented member.

$sel:status:ActionHistory', actionHistory_status - The status of action at the time of the event.

$sel:eventType:ActionHistory', actionHistory_eventType - This distinguishes between whether the events are triggered by the user or generated by the system.

$sel:actionHistoryDetails:ActionHistory', actionHistory_actionHistoryDetails - The description of details of the event.

actionHistory_timestamp :: Lens' ActionHistory UTCTime

Undocumented member.

actionHistory_status :: Lens' ActionHistory ActionStatus

The status of action at the time of the event.

actionHistory_eventType :: Lens' ActionHistory EventType

This distinguishes between whether the events are triggered by the user or generated by the system.

actionHistory_actionHistoryDetails :: Lens' ActionHistory ActionHistoryDetails

The description of details of the event.

ActionHistoryDetails

data ActionHistoryDetails

The description of details of the event.

See: newActionHistoryDetails smart constructor.

Constructors

ActionHistoryDetails'

Fields

message :: Text

action :: Action

The budget action resource.

newActionHistoryDetails

Arguments

:: Text
$sel:message:ActionHistoryDetails'

-> Action
$sel:action:ActionHistoryDetails'

-> ActionHistoryDetails

Create a value of ActionHistoryDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:ActionHistoryDetails', actionHistoryDetails_message - Undocumented member.

$sel:action:ActionHistoryDetails', actionHistoryDetails_action - The budget action resource.

actionHistoryDetails_message :: Lens' ActionHistoryDetails Text

Undocumented member.

actionHistoryDetails_action :: Lens' ActionHistoryDetails Action

The budget action resource.

ActionThreshold

data ActionThreshold

The trigger threshold of the action.

See: newActionThreshold smart constructor.

Constructors

ActionThreshold'

Fields

actionThresholdValue :: Double

actionThresholdType :: ThresholdType

newActionThreshold

Arguments

:: Double
$sel:actionThresholdValue:ActionThreshold'

-> ThresholdType
$sel:actionThresholdType:ActionThreshold'

-> ActionThreshold

Create a value of ActionThreshold with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionThresholdValue:ActionThreshold', actionThreshold_actionThresholdValue - Undocumented member.

$sel:actionThresholdType:ActionThreshold', actionThreshold_actionThresholdType - Undocumented member.

actionThreshold_actionThresholdValue :: Lens' ActionThreshold Double

Undocumented member.

actionThreshold_actionThresholdType :: Lens' ActionThreshold ThresholdType

Undocumented member.

Budget

data Budget

Represents the output of the CreateBudget operation. The content consists of the detailed metadata and data file information, and the current status of the budget object.

This is the ARN pattern for a budget:

arn:aws:budgets::AccountId:budget/budgetName

See: newBudget smart constructor.

Constructors

Budget'

Fields

calculatedSpend :: Maybe CalculatedSpend

The actual and forecasted cost or usage that the budget tracks.

plannedBudgetLimits :: Maybe (HashMap Text Spend)

A map containing multiple BudgetLimit, including current or future limits.

PlannedBudgetLimits is available for cost or usage budget and supports monthly and quarterly TimeUnit.

For monthly budgets, provide 12 months of PlannedBudgetLimits values. This must start from the current month and include the next 11 months. The key is the start of the month, UTC in epoch seconds.

For quarterly budgets, provide 4 quarters of PlannedBudgetLimits value entries in standard calendar quarter increments. This must start from the current quarter and include the next 3 quarters. The key is the start of the quarter, UTC in epoch seconds.

If the planned budget expires before 12 months for monthly or 4 quarters for quarterly, provide the PlannedBudgetLimits values only for the remaining periods.

If the budget begins at a date in the future, provide PlannedBudgetLimits values from the start date of the budget.

After all of the BudgetLimit values in PlannedBudgetLimits are used, the budget continues to use the last limit as the BudgetLimit. At that point, the planned budget provides the same experience as a fixed budget.

DescribeBudget and DescribeBudgets response along with PlannedBudgetLimits will also contain BudgetLimit representing the current month or quarter limit present in PlannedBudgetLimits. This only applies to budgets created with PlannedBudgetLimits. Budgets created without PlannedBudgetLimits will only contain BudgetLimit, and no PlannedBudgetLimits.

lastUpdatedTime :: Maybe POSIX

The last time that you updated this budget.

budgetLimit :: Maybe Spend

The total amount of cost, usage, RI utilization, RI coverage, Savings Plans utilization, or Savings Plans coverage that you want to track with your budget.

BudgetLimit is required for cost or usage budgets, but optional for RI or Savings Plans utilization or coverage budgets. RI and Savings Plans utilization or coverage budgets default to 100, which is the only valid value for RI or Savings Plans utilization or coverage budgets. You can't use BudgetLimit with PlannedBudgetLimits for CreateBudget and UpdateBudget actions.

timePeriod :: Maybe TimePeriod

The period of time that is covered by a budget. The period has a start date and an end date. The start date must come before the end date. The end date must come before 06/15/87 00:00 UTC.

If you create your budget and don't specify a start date, AWS defaults to the start of your chosen time period (DAILY, MONTHLY, QUARTERLY, or ANNUALLY). For example, if you created your budget on January 24, 2018, chose DAILY, and didn't set a start date, AWS set your start date to 01/24/18 00:00 UTC. If you chose MONTHLY, AWS set your start date to 01/01/18 00:00 UTC. If you didn't specify an end date, AWS set your end date to 06/15/87 00:00 UTC. The defaults are the same for the AWS Billing and Cost Management console and the API.

You can change either date with the UpdateBudget operation.

After the end date, AWS deletes the budget and all associated notifications and subscribers.

costTypes :: Maybe CostTypes

The types of costs that are included in this COST budget.

USAGE, RI_UTILIZATION, RI_COVERAGE, SAVINGS_PLANS_UTILIZATION, and SAVINGS_PLANS_COVERAGE budgets do not have CostTypes.

costFilters :: Maybe (HashMap Text [Text])

The cost filters, such as service or tag, that are applied to a budget.

AWS Budgets supports the following services as a filter for RI budgets:

Amazon Elastic Compute Cloud - Compute

Amazon Redshift

Amazon Relational Database Service

Amazon ElastiCache

Amazon Elasticsearch Service

budgetName :: Text

The name of a budget. The name must be unique within an account. The : and \ characters aren't allowed in BudgetName.

timeUnit :: TimeUnit

The length of time until a budget resets the actual and forecasted spend.

budgetType :: BudgetType

Whether this budget tracks costs, usage, RI utilization, RI coverage, Savings Plans utilization, or Savings Plans coverage.

newBudget

Arguments

:: Text
$sel:budgetName:Budget'

-> TimeUnit
$sel:timeUnit:Budget'

-> BudgetType
$sel:budgetType:Budget'

-> Budget

Create a value of Budget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:calculatedSpend:Budget', budget_calculatedSpend - The actual and forecasted cost or usage that the budget tracks.

$sel:plannedBudgetLimits:Budget', budget_plannedBudgetLimits - A map containing multiple BudgetLimit, including current or future limits.

PlannedBudgetLimits is available for cost or usage budget and supports monthly and quarterly TimeUnit.

For monthly budgets, provide 12 months of PlannedBudgetLimits values. This must start from the current month and include the next 11 months. The key is the start of the month, UTC in epoch seconds.

For quarterly budgets, provide 4 quarters of PlannedBudgetLimits value entries in standard calendar quarter increments. This must start from the current quarter and include the next 3 quarters. The key is the start of the quarter, UTC in epoch seconds.

If the planned budget expires before 12 months for monthly or 4 quarters for quarterly, provide the PlannedBudgetLimits values only for the remaining periods.

If the budget begins at a date in the future, provide PlannedBudgetLimits values from the start date of the budget.

After all of the BudgetLimit values in PlannedBudgetLimits are used, the budget continues to use the last limit as the BudgetLimit. At that point, the planned budget provides the same experience as a fixed budget.

DescribeBudget and DescribeBudgets response along with PlannedBudgetLimits will also contain BudgetLimit representing the current month or quarter limit present in PlannedBudgetLimits. This only applies to budgets created with PlannedBudgetLimits. Budgets created without PlannedBudgetLimits will only contain BudgetLimit, and no PlannedBudgetLimits.

$sel:lastUpdatedTime:Budget', budget_lastUpdatedTime - The last time that you updated this budget.

$sel:budgetLimit:Budget', budget_budgetLimit - The total amount of cost, usage, RI utilization, RI coverage, Savings Plans utilization, or Savings Plans coverage that you want to track with your budget.

BudgetLimit is required for cost or usage budgets, but optional for RI or Savings Plans utilization or coverage budgets. RI and Savings Plans utilization or coverage budgets default to 100, which is the only valid value for RI or Savings Plans utilization or coverage budgets. You can't use BudgetLimit with PlannedBudgetLimits for CreateBudget and UpdateBudget actions.

$sel:timePeriod:Budget', budget_timePeriod - The period of time that is covered by a budget. The period has a start date and an end date. The start date must come before the end date. The end date must come before 06/15/87 00:00 UTC.

If you create your budget and don't specify a start date, AWS defaults to the start of your chosen time period (DAILY, MONTHLY, QUARTERLY, or ANNUALLY). For example, if you created your budget on January 24, 2018, chose DAILY, and didn't set a start date, AWS set your start date to 01/24/18 00:00 UTC. If you chose MONTHLY, AWS set your start date to 01/01/18 00:00 UTC. If you didn't specify an end date, AWS set your end date to 06/15/87 00:00 UTC. The defaults are the same for the AWS Billing and Cost Management console and the API.

You can change either date with the UpdateBudget operation.

After the end date, AWS deletes the budget and all associated notifications and subscribers.

$sel:costTypes:Budget', budget_costTypes - The types of costs that are included in this COST budget.

USAGE, RI_UTILIZATION, RI_COVERAGE, SAVINGS_PLANS_UTILIZATION, and SAVINGS_PLANS_COVERAGE budgets do not have CostTypes.

$sel:costFilters:Budget', budget_costFilters - The cost filters, such as service or tag, that are applied to a budget.

AWS Budgets supports the following services as a filter for RI budgets:

Amazon Elastic Compute Cloud - Compute

Amazon Redshift

Amazon Relational Database Service

Amazon ElastiCache

Amazon Elasticsearch Service

$sel:budgetName:Budget', budget_budgetName - The name of a budget. The name must be unique within an account. The : and \ characters aren't allowed in BudgetName.

$sel:timeUnit:Budget', budget_timeUnit - The length of time until a budget resets the actual and forecasted spend.

$sel:budgetType:Budget', budget_budgetType - Whether this budget tracks costs, usage, RI utilization, RI coverage, Savings Plans utilization, or Savings Plans coverage.

budget_calculatedSpend :: Lens' Budget (Maybe CalculatedSpend)

The actual and forecasted cost or usage that the budget tracks.

budget_plannedBudgetLimits :: Lens' Budget (Maybe (HashMap Text Spend))

A map containing multiple BudgetLimit, including current or future limits.

PlannedBudgetLimits is available for cost or usage budget and supports monthly and quarterly TimeUnit.

For monthly budgets, provide 12 months of PlannedBudgetLimits values. This must start from the current month and include the next 11 months. The key is the start of the month, UTC in epoch seconds.

For quarterly budgets, provide 4 quarters of PlannedBudgetLimits value entries in standard calendar quarter increments. This must start from the current quarter and include the next 3 quarters. The key is the start of the quarter, UTC in epoch seconds.

If the planned budget expires before 12 months for monthly or 4 quarters for quarterly, provide the PlannedBudgetLimits values only for the remaining periods.

If the budget begins at a date in the future, provide PlannedBudgetLimits values from the start date of the budget.

After all of the BudgetLimit values in PlannedBudgetLimits are used, the budget continues to use the last limit as the BudgetLimit. At that point, the planned budget provides the same experience as a fixed budget.

DescribeBudget and DescribeBudgets response along with PlannedBudgetLimits will also contain BudgetLimit representing the current month or quarter limit present in PlannedBudgetLimits. This only applies to budgets created with PlannedBudgetLimits. Budgets created without PlannedBudgetLimits will only contain BudgetLimit, and no PlannedBudgetLimits.

budget_lastUpdatedTime :: Lens' Budget (Maybe UTCTime)

The last time that you updated this budget.

budget_budgetLimit :: Lens' Budget (Maybe Spend)

The total amount of cost, usage, RI utilization, RI coverage, Savings Plans utilization, or Savings Plans coverage that you want to track with your budget.

BudgetLimit is required for cost or usage budgets, but optional for RI or Savings Plans utilization or coverage budgets. RI and Savings Plans utilization or coverage budgets default to 100, which is the only valid value for RI or Savings Plans utilization or coverage budgets. You can't use BudgetLimit with PlannedBudgetLimits for CreateBudget and UpdateBudget actions.

budget_timePeriod :: Lens' Budget (Maybe TimePeriod)

The period of time that is covered by a budget. The period has a start date and an end date. The start date must come before the end date. The end date must come before 06/15/87 00:00 UTC.

If you create your budget and don't specify a start date, AWS defaults to the start of your chosen time period (DAILY, MONTHLY, QUARTERLY, or ANNUALLY). For example, if you created your budget on January 24, 2018, chose DAILY, and didn't set a start date, AWS set your start date to 01/24/18 00:00 UTC. If you chose MONTHLY, AWS set your start date to 01/01/18 00:00 UTC. If you didn't specify an end date, AWS set your end date to 06/15/87 00:00 UTC. The defaults are the same for the AWS Billing and Cost Management console and the API.

You can change either date with the UpdateBudget operation.

After the end date, AWS deletes the budget and all associated notifications and subscribers.

budget_costTypes :: Lens' Budget (Maybe CostTypes)

The types of costs that are included in this COST budget.

USAGE, RI_UTILIZATION, RI_COVERAGE, SAVINGS_PLANS_UTILIZATION, and SAVINGS_PLANS_COVERAGE budgets do not have CostTypes.

budget_costFilters :: Lens' Budget (Maybe (HashMap Text [Text]))

The cost filters, such as service or tag, that are applied to a budget.

AWS Budgets supports the following services as a filter for RI budgets:

Amazon Elastic Compute Cloud - Compute

Amazon Redshift

Amazon Relational Database Service

Amazon ElastiCache

Amazon Elasticsearch Service

budget_budgetName :: Lens' Budget Text

The name of a budget. The name must be unique within an account. The : and \ characters aren't allowed in BudgetName.

budget_timeUnit :: Lens' Budget TimeUnit

The length of time until a budget resets the actual and forecasted spend.

budget_budgetType :: Lens' Budget BudgetType

Whether this budget tracks costs, usage, RI utilization, RI coverage, Savings Plans utilization, or Savings Plans coverage.

BudgetPerformanceHistory

data BudgetPerformanceHistory

A history of the state of a budget at the end of the budget's specified time period.

See: newBudgetPerformanceHistory smart constructor.

Constructors

BudgetPerformanceHistory'

Fields

budgetedAndActualAmountsList :: Maybe [BudgetedAndActualAmounts]

A list of amounts of cost or usage that you created budgets for, compared to your actual costs or usage.

timeUnit :: Maybe TimeUnit

budgetName :: Maybe Text

budgetType :: Maybe BudgetType

costTypes :: Maybe CostTypes

The history of the cost types for a budget during the specified time period.

costFilters :: Maybe (HashMap Text [Text])

The history of the cost filters for a budget during the specified time period.

newBudgetPerformanceHistory :: BudgetPerformanceHistory

Create a value of BudgetPerformanceHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:budgetedAndActualAmountsList:BudgetPerformanceHistory', budgetPerformanceHistory_budgetedAndActualAmountsList - A list of amounts of cost or usage that you created budgets for, compared to your actual costs or usage.

$sel:timeUnit:BudgetPerformanceHistory', budgetPerformanceHistory_timeUnit - Undocumented member.

$sel:budgetName:BudgetPerformanceHistory', budgetPerformanceHistory_budgetName - Undocumented member.

$sel:budgetType:BudgetPerformanceHistory', budgetPerformanceHistory_budgetType - Undocumented member.

$sel:costTypes:BudgetPerformanceHistory', budgetPerformanceHistory_costTypes - The history of the cost types for a budget during the specified time period.

$sel:costFilters:BudgetPerformanceHistory', budgetPerformanceHistory_costFilters - The history of the cost filters for a budget during the specified time period.

budgetPerformanceHistory_budgetedAndActualAmountsList :: Lens' BudgetPerformanceHistory (Maybe [BudgetedAndActualAmounts])

A list of amounts of cost or usage that you created budgets for, compared to your actual costs or usage.

budgetPerformanceHistory_timeUnit :: Lens' BudgetPerformanceHistory (Maybe TimeUnit)

Undocumented member.

budgetPerformanceHistory_budgetName :: Lens' BudgetPerformanceHistory (Maybe Text)

Undocumented member.

budgetPerformanceHistory_budgetType :: Lens' BudgetPerformanceHistory (Maybe BudgetType)

Undocumented member.

budgetPerformanceHistory_costTypes :: Lens' BudgetPerformanceHistory (Maybe CostTypes)

The history of the cost types for a budget during the specified time period.

budgetPerformanceHistory_costFilters :: Lens' BudgetPerformanceHistory (Maybe (HashMap Text [Text]))

The history of the cost filters for a budget during the specified time period.

BudgetedAndActualAmounts

data BudgetedAndActualAmounts

The amount of cost or usage that you created the budget for, compared to your actual costs or usage.

See: newBudgetedAndActualAmounts smart constructor.

Constructors

BudgetedAndActualAmounts'

Fields

timePeriod :: Maybe TimePeriod

The time period covered by this budget comparison.

actualAmount :: Maybe Spend

Your actual costs or usage for a budget period.

budgetedAmount :: Maybe Spend

The amount of cost or usage that you created the budget for.

newBudgetedAndActualAmounts :: BudgetedAndActualAmounts

Create a value of BudgetedAndActualAmounts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timePeriod:BudgetedAndActualAmounts', budgetedAndActualAmounts_timePeriod - The time period covered by this budget comparison.

$sel:actualAmount:BudgetedAndActualAmounts', budgetedAndActualAmounts_actualAmount - Your actual costs or usage for a budget period.

$sel:budgetedAmount:BudgetedAndActualAmounts', budgetedAndActualAmounts_budgetedAmount - The amount of cost or usage that you created the budget for.

budgetedAndActualAmounts_timePeriod :: Lens' BudgetedAndActualAmounts (Maybe TimePeriod)

The time period covered by this budget comparison.

budgetedAndActualAmounts_actualAmount :: Lens' BudgetedAndActualAmounts (Maybe Spend)

Your actual costs or usage for a budget period.

budgetedAndActualAmounts_budgetedAmount :: Lens' BudgetedAndActualAmounts (Maybe Spend)

The amount of cost or usage that you created the budget for.

CalculatedSpend

data CalculatedSpend

The spend objects that are associated with this budget. The actualSpend tracks how much you've used, cost, usage, RI units, or Savings Plans units and the forecastedSpend tracks how much you are predicted to spend based on your historical usage profile.

For example, if it is the 20th of the month and you have spent 50 dollars on Amazon EC2, your actualSpend is 50 USD, and your forecastedSpend is 75 USD.

See: newCalculatedSpend smart constructor.

Constructors

CalculatedSpend'

Fields

forecastedSpend :: Maybe Spend

The amount of cost, usage, RI units, or Savings Plans units that you are forecasted to use.

actualSpend :: Spend

The amount of cost, usage, RI units, or Savings Plans units that you have used.

newCalculatedSpend

Arguments

:: Spend
$sel:actualSpend:CalculatedSpend'

-> CalculatedSpend

Create a value of CalculatedSpend with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:forecastedSpend:CalculatedSpend', calculatedSpend_forecastedSpend - The amount of cost, usage, RI units, or Savings Plans units that you are forecasted to use.

$sel:actualSpend:CalculatedSpend', calculatedSpend_actualSpend - The amount of cost, usage, RI units, or Savings Plans units that you have used.

calculatedSpend_forecastedSpend :: Lens' CalculatedSpend (Maybe Spend)

The amount of cost, usage, RI units, or Savings Plans units that you are forecasted to use.

calculatedSpend_actualSpend :: Lens' CalculatedSpend Spend

The amount of cost, usage, RI units, or Savings Plans units that you have used.

CostTypes

data CostTypes

The types of cost that are included in a COST budget, such as tax and subscriptions.

USAGE, RI_UTILIZATION, RI_COVERAGE, SAVINGS_PLANS_UTILIZATION, and SAVINGS_PLANS_COVERAGE budgets do not have CostTypes.

See: newCostTypes smart constructor.

Constructors

CostTypes'

Fields

useAmortized :: Maybe Bool

Specifies whether a budget uses the amortized rate.

The default value is false.

includeRecurring :: Maybe Bool

Specifies whether a budget includes recurring fees such as monthly RI fees.

The default value is true.

useBlended :: Maybe Bool

Specifies whether a budget uses a blended rate.

The default value is false.

includeSupport :: Maybe Bool

Specifies whether a budget includes support subscription fees.

The default value is true.

includeDiscount :: Maybe Bool

Specifies whether a budget includes discounts.

The default value is true.

includeSubscription :: Maybe Bool

Specifies whether a budget includes subscriptions.

The default value is true.

includeRefund :: Maybe Bool

Specifies whether a budget includes refunds.

The default value is true.

includeUpfront :: Maybe Bool

Specifies whether a budget includes upfront RI costs.

The default value is true.

includeOtherSubscription :: Maybe Bool

Specifies whether a budget includes non-RI subscription costs.

The default value is true.

includeTax :: Maybe Bool

Specifies whether a budget includes taxes.

The default value is true.

includeCredit :: Maybe Bool

Specifies whether a budget includes credits.

The default value is true.

newCostTypes :: CostTypes

Create a value of CostTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:useAmortized:CostTypes', costTypes_useAmortized - Specifies whether a budget uses the amortized rate.

The default value is false.

$sel:includeRecurring:CostTypes', costTypes_includeRecurring - Specifies whether a budget includes recurring fees such as monthly RI fees.

The default value is true.

$sel:useBlended:CostTypes', costTypes_useBlended - Specifies whether a budget uses a blended rate.

The default value is false.

$sel:includeSupport:CostTypes', costTypes_includeSupport - Specifies whether a budget includes support subscription fees.

The default value is true.

$sel:includeDiscount:CostTypes', costTypes_includeDiscount - Specifies whether a budget includes discounts.

The default value is true.

$sel:includeSubscription:CostTypes', costTypes_includeSubscription - Specifies whether a budget includes subscriptions.

The default value is true.

$sel:includeRefund:CostTypes', costTypes_includeRefund - Specifies whether a budget includes refunds.

The default value is true.

$sel:includeUpfront:CostTypes', costTypes_includeUpfront - Specifies whether a budget includes upfront RI costs.

The default value is true.

$sel:includeOtherSubscription:CostTypes', costTypes_includeOtherSubscription - Specifies whether a budget includes non-RI subscription costs.

The default value is true.

$sel:includeTax:CostTypes', costTypes_includeTax - Specifies whether a budget includes taxes.

The default value is true.

$sel:includeCredit:CostTypes', costTypes_includeCredit - Specifies whether a budget includes credits.

The default value is true.

costTypes_useAmortized :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget uses the amortized rate.

The default value is false.

costTypes_includeRecurring :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes recurring fees such as monthly RI fees.

The default value is true.

costTypes_useBlended :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget uses a blended rate.

The default value is false.

costTypes_includeSupport :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes support subscription fees.

The default value is true.

costTypes_includeDiscount :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes discounts.

The default value is true.

costTypes_includeSubscription :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes subscriptions.

The default value is true.

costTypes_includeRefund :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes refunds.

The default value is true.

costTypes_includeUpfront :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes upfront RI costs.

The default value is true.

costTypes_includeOtherSubscription :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes non-RI subscription costs.

The default value is true.

costTypes_includeTax :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes taxes.

The default value is true.

costTypes_includeCredit :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes credits.

The default value is true.

Definition

data Definition

Specifies all of the type-specific parameters.

See: newDefinition smart constructor.

Constructors

Definition'

Fields

scpActionDefinition :: Maybe ScpActionDefinition

The service control policies (SCPs) action definition details.

iamActionDefinition :: Maybe IamActionDefinition

The AWS Identity and Access Management (IAM) action definition details.

ssmActionDefinition :: Maybe SsmActionDefinition

The AWS Systems Manager (SSM) action definition details.

newDefinition :: Definition

Create a value of Definition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scpActionDefinition:Definition', definition_scpActionDefinition - The service control policies (SCPs) action definition details.

$sel:iamActionDefinition:Definition', definition_iamActionDefinition - The AWS Identity and Access Management (IAM) action definition details.

$sel:ssmActionDefinition:Definition', definition_ssmActionDefinition - The AWS Systems Manager (SSM) action definition details.

definition_scpActionDefinition :: Lens' Definition (Maybe ScpActionDefinition)

The service control policies (SCPs) action definition details.

definition_iamActionDefinition :: Lens' Definition (Maybe IamActionDefinition)

The AWS Identity and Access Management (IAM) action definition details.

definition_ssmActionDefinition :: Lens' Definition (Maybe SsmActionDefinition)

The AWS Systems Manager (SSM) action definition details.

IamActionDefinition

data IamActionDefinition

The AWS Identity and Access Management (IAM) action definition details.

See: newIamActionDefinition smart constructor.

Constructors

IamActionDefinition'

Fields

groups :: Maybe (NonEmpty Text)

A list of groups to be attached. There must be at least one group.

roles :: Maybe (NonEmpty Text)

A list of roles to be attached. There must be at least one role.

users :: Maybe (NonEmpty Text)

A list of users to be attached. There must be at least one user.

policyArn :: Text

The Amazon Resource Name (ARN) of the policy to be attached.

newIamActionDefinition

Arguments

:: Text
$sel:policyArn:IamActionDefinition'

-> IamActionDefinition

Create a value of IamActionDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:IamActionDefinition', iamActionDefinition_groups - A list of groups to be attached. There must be at least one group.

$sel:roles:IamActionDefinition', iamActionDefinition_roles - A list of roles to be attached. There must be at least one role.

$sel:users:IamActionDefinition', iamActionDefinition_users - A list of users to be attached. There must be at least one user.

$sel:policyArn:IamActionDefinition', iamActionDefinition_policyArn - The Amazon Resource Name (ARN) of the policy to be attached.

iamActionDefinition_groups :: Lens' IamActionDefinition (Maybe (NonEmpty Text))

A list of groups to be attached. There must be at least one group.

iamActionDefinition_roles :: Lens' IamActionDefinition (Maybe (NonEmpty Text))

A list of roles to be attached. There must be at least one role.

iamActionDefinition_users :: Lens' IamActionDefinition (Maybe (NonEmpty Text))

A list of users to be attached. There must be at least one user.

iamActionDefinition_policyArn :: Lens' IamActionDefinition Text

The Amazon Resource Name (ARN) of the policy to be attached.

Notification

data Notification

A notification that is associated with a budget. A budget can have up to ten notifications.

Each notification must have at least one subscriber. A notification can have one SNS subscriber and up to 10 email subscribers, for a total of 11 subscribers.

For example, if you have a budget for 200 dollars and you want to be notified when you go over 160 dollars, create a notification with the following parameters:

A notificationType of ACTUAL

A thresholdType of PERCENTAGE

A comparisonOperator of GREATER_THAN

A notification threshold of 80

See: newNotification smart constructor.

Constructors

Notification'

Fields

thresholdType :: Maybe ThresholdType

The type of threshold for a notification. For ABSOLUTE_VALUE thresholds, AWS notifies you when you go over or are forecasted to go over your total cost threshold. For PERCENTAGE thresholds, AWS notifies you when you go over or are forecasted to go over a certain percentage of your forecasted spend. For example, if you have a budget for 200 dollars and you have a PERCENTAGE threshold of 80%, AWS notifies you when you go over 160 dollars.

notificationState :: Maybe NotificationState

Whether this notification is in alarm. If a budget notification is in the ALARM state, you have passed the set threshold for the budget.

notificationType :: NotificationType

Whether the notification is for how much you have spent (ACTUAL) or for how much you're forecasted to spend (FORECASTED).

comparisonOperator :: ComparisonOperator

The comparison that is used for this notification.

threshold :: Double

The threshold that is associated with a notification. Thresholds are always a percentage, and many customers find value being alerted between 50% - 200% of the budgeted amount. The maximum limit for your threshold is 1,000,000% above the budgeted amount.

newNotification

Arguments

:: NotificationType
$sel:notificationType:Notification'

-> ComparisonOperator
$sel:comparisonOperator:Notification'

-> Double
$sel:threshold:Notification'

-> Notification

Create a value of Notification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thresholdType:Notification', notification_thresholdType - The type of threshold for a notification. For ABSOLUTE_VALUE thresholds, AWS notifies you when you go over or are forecasted to go over your total cost threshold. For PERCENTAGE thresholds, AWS notifies you when you go over or are forecasted to go over a certain percentage of your forecasted spend. For example, if you have a budget for 200 dollars and you have a PERCENTAGE threshold of 80%, AWS notifies you when you go over 160 dollars.

$sel:notificationState:Notification', notification_notificationState - Whether this notification is in alarm. If a budget notification is in the ALARM state, you have passed the set threshold for the budget.

$sel:notificationType:Notification', notification_notificationType - Whether the notification is for how much you have spent (ACTUAL) or for how much you're forecasted to spend (FORECASTED).

$sel:comparisonOperator:Notification', notification_comparisonOperator - The comparison that is used for this notification.

$sel:threshold:Notification', notification_threshold - The threshold that is associated with a notification. Thresholds are always a percentage, and many customers find value being alerted between 50% - 200% of the budgeted amount. The maximum limit for your threshold is 1,000,000% above the budgeted amount.

notification_thresholdType :: Lens' Notification (Maybe ThresholdType)

The type of threshold for a notification. For ABSOLUTE_VALUE thresholds, AWS notifies you when you go over or are forecasted to go over your total cost threshold. For PERCENTAGE thresholds, AWS notifies you when you go over or are forecasted to go over a certain percentage of your forecasted spend. For example, if you have a budget for 200 dollars and you have a PERCENTAGE threshold of 80%, AWS notifies you when you go over 160 dollars.

notification_notificationState :: Lens' Notification (Maybe NotificationState)

Whether this notification is in alarm. If a budget notification is in the ALARM state, you have passed the set threshold for the budget.

notification_notificationType :: Lens' Notification NotificationType

Whether the notification is for how much you have spent (ACTUAL) or for how much you're forecasted to spend (FORECASTED).

notification_comparisonOperator :: Lens' Notification ComparisonOperator

The comparison that is used for this notification.

notification_threshold :: Lens' Notification Double

The threshold that is associated with a notification. Thresholds are always a percentage, and many customers find value being alerted between 50% - 200% of the budgeted amount. The maximum limit for your threshold is 1,000,000% above the budgeted amount.

NotificationWithSubscribers

data NotificationWithSubscribers

A notification with subscribers. A notification can have one SNS subscriber and up to 10 email subscribers, for a total of 11 subscribers.

See: newNotificationWithSubscribers smart constructor.

Constructors

NotificationWithSubscribers'

Fields

notification :: Notification

The notification that is associated with a budget.

subscribers :: NonEmpty Subscriber

A list of subscribers who are subscribed to this notification.

newNotificationWithSubscribers

Arguments

:: Notification
$sel:notification:NotificationWithSubscribers'

-> NonEmpty Subscriber
$sel:subscribers:NotificationWithSubscribers'

-> NotificationWithSubscribers

Create a value of NotificationWithSubscribers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notification:NotificationWithSubscribers', notificationWithSubscribers_notification - The notification that is associated with a budget.

$sel:subscribers:NotificationWithSubscribers', notificationWithSubscribers_subscribers - A list of subscribers who are subscribed to this notification.

notificationWithSubscribers_notification :: Lens' NotificationWithSubscribers Notification

The notification that is associated with a budget.

notificationWithSubscribers_subscribers :: Lens' NotificationWithSubscribers (NonEmpty Subscriber)

A list of subscribers who are subscribed to this notification.

ScpActionDefinition

data ScpActionDefinition

The service control policies (SCP) action definition details.

See: newScpActionDefinition smart constructor.

Constructors

ScpActionDefinition'

Fields

policyId :: Text

The policy ID attached.

targetIds :: NonEmpty Text

A list of target IDs.

newScpActionDefinition

Arguments

:: Text
$sel:policyId:ScpActionDefinition'

-> NonEmpty Text
$sel:targetIds:ScpActionDefinition'

-> ScpActionDefinition

Create a value of ScpActionDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyId:ScpActionDefinition', scpActionDefinition_policyId - The policy ID attached.

$sel:targetIds:ScpActionDefinition', scpActionDefinition_targetIds - A list of target IDs.

scpActionDefinition_policyId :: Lens' ScpActionDefinition Text

The policy ID attached.

scpActionDefinition_targetIds :: Lens' ScpActionDefinition (NonEmpty Text)

A list of target IDs.

Spend

data Spend

The amount of cost or usage that is measured for a budget.

For example, a Spend for 3 GB of S3 usage would have the following parameters:

An Amount of 3

A unit of GB

See: newSpend smart constructor.

Constructors

Spend'

Fields

amount :: Text

The cost or usage amount that is associated with a budget forecast, actual spend, or budget threshold.

unit :: Text

The unit of measurement that is used for the budget forecast, actual spend, or budget threshold, such as dollars or GB.

newSpend

Arguments

:: Text
$sel:amount:Spend'

-> Text
$sel:unit:Spend'

-> Spend

Create a value of Spend with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amount:Spend', spend_amount - The cost or usage amount that is associated with a budget forecast, actual spend, or budget threshold.

$sel:unit:Spend', spend_unit - The unit of measurement that is used for the budget forecast, actual spend, or budget threshold, such as dollars or GB.

spend_amount :: Lens' Spend Text

The cost or usage amount that is associated with a budget forecast, actual spend, or budget threshold.

spend_unit :: Lens' Spend Text

The unit of measurement that is used for the budget forecast, actual spend, or budget threshold, such as dollars or GB.

SsmActionDefinition

data SsmActionDefinition

The AWS Systems Manager (SSM) action definition details.

See: newSsmActionDefinition smart constructor.

Constructors

SsmActionDefinition'

Fields

actionSubType :: ActionSubType

The action subType.

region :: Text

The Region to run the SSM document.

instanceIds :: NonEmpty Text

The EC2 and RDS instance IDs.

newSsmActionDefinition

Arguments

:: ActionSubType
$sel:actionSubType:SsmActionDefinition'

-> Text
$sel:region:SsmActionDefinition'

-> NonEmpty Text
$sel:instanceIds:SsmActionDefinition'

-> SsmActionDefinition

Create a value of SsmActionDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionSubType:SsmActionDefinition', ssmActionDefinition_actionSubType - The action subType.

$sel:region:SsmActionDefinition', ssmActionDefinition_region - The Region to run the SSM document.

$sel:instanceIds:SsmActionDefinition', ssmActionDefinition_instanceIds - The EC2 and RDS instance IDs.

ssmActionDefinition_actionSubType :: Lens' SsmActionDefinition ActionSubType

The action subType.

ssmActionDefinition_region :: Lens' SsmActionDefinition Text

The Region to run the SSM document.

ssmActionDefinition_instanceIds :: Lens' SsmActionDefinition (NonEmpty Text)

The EC2 and RDS instance IDs.

Subscriber

data Subscriber

The subscriber to a budget notification. The subscriber consists of a subscription type and either an Amazon SNS topic or an email address.

For example, an email subscriber would have the following parameters:

A subscriptionType of EMAIL

An address of example@example.com

See: newSubscriber smart constructor.

Constructors

Subscriber'

Fields

subscriptionType :: SubscriptionType

The type of notification that AWS sends to a subscriber.

address :: Sensitive Text

The address that AWS sends budget notifications to, either an SNS topic or an email.

When you create a subscriber, the value of Address can't contain line breaks.

newSubscriber

Arguments

:: SubscriptionType
$sel:subscriptionType:Subscriber'

-> Text
$sel:address:Subscriber'

-> Subscriber

Create a value of Subscriber with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionType:Subscriber', subscriber_subscriptionType - The type of notification that AWS sends to a subscriber.

$sel:address:Subscriber', subscriber_address - The address that AWS sends budget notifications to, either an SNS topic or an email.

When you create a subscriber, the value of Address can't contain line breaks.

subscriber_subscriptionType :: Lens' Subscriber SubscriptionType

The type of notification that AWS sends to a subscriber.

subscriber_address :: Lens' Subscriber Text

The address that AWS sends budget notifications to, either an SNS topic or an email.

When you create a subscriber, the value of Address can't contain line breaks.

TimePeriod

data TimePeriod

The period of time that is covered by a budget. The period has a start date and an end date. The start date must come before the end date. There are no restrictions on the end date.

See: newTimePeriod smart constructor.

Constructors

TimePeriod'

Fields

start :: Maybe POSIX

The start date for a budget. If you created your budget and didn't specify a start date, AWS defaults to the start of your chosen time period (DAILY, MONTHLY, QUARTERLY, or ANNUALLY). For example, if you created your budget on January 24, 2018, chose DAILY, and didn't set a start date, AWS set your start date to 01/24/18 00:00 UTC. If you chose MONTHLY, AWS set your start date to 01/01/18 00:00 UTC. The defaults are the same for the AWS Billing and Cost Management console and the API.

You can change your start date with the UpdateBudget operation.

end :: Maybe POSIX

The end date for a budget. If you didn't specify an end date, AWS set your end date to 06/15/87 00:00 UTC. The defaults are the same for the AWS Billing and Cost Management console and the API.

After the end date, AWS deletes the budget and all associated notifications and subscribers. You can change your end date with the UpdateBudget operation.

newTimePeriod :: TimePeriod

Create a value of TimePeriod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:start:TimePeriod', timePeriod_start - The start date for a budget. If you created your budget and didn't specify a start date, AWS defaults to the start of your chosen time period (DAILY, MONTHLY, QUARTERLY, or ANNUALLY). For example, if you created your budget on January 24, 2018, chose DAILY, and didn't set a start date, AWS set your start date to 01/24/18 00:00 UTC. If you chose MONTHLY, AWS set your start date to 01/01/18 00:00 UTC. The defaults are the same for the AWS Billing and Cost Management console and the API.

You can change your start date with the UpdateBudget operation.

$sel:end:TimePeriod', timePeriod_end - The end date for a budget. If you didn't specify an end date, AWS set your end date to 06/15/87 00:00 UTC. The defaults are the same for the AWS Billing and Cost Management console and the API.

After the end date, AWS deletes the budget and all associated notifications and subscribers. You can change your end date with the UpdateBudget operation.

timePeriod_start :: Lens' TimePeriod (Maybe UTCTime)

The start date for a budget. If you created your budget and didn't specify a start date, AWS defaults to the start of your chosen time period (DAILY, MONTHLY, QUARTERLY, or ANNUALLY). For example, if you created your budget on January 24, 2018, chose DAILY, and didn't set a start date, AWS set your start date to 01/24/18 00:00 UTC. If you chose MONTHLY, AWS set your start date to 01/01/18 00:00 UTC. The defaults are the same for the AWS Billing and Cost Management console and the API.

You can change your start date with the UpdateBudget operation.

timePeriod_end :: Lens' TimePeriod (Maybe UTCTime)

The end date for a budget. If you didn't specify an end date, AWS set your end date to 06/15/87 00:00 UTC. The defaults are the same for the AWS Billing and Cost Management console and the API.

After the end date, AWS deletes the budget and all associated notifications and subscribers. You can change your end date with the UpdateBudget operation.