Network.AWS.Budgets.Types.Action

Description

Documentation

data Action

A budget action resource.

See: newAction smart constructor.

Constructors

Action'

Fields

actionId :: Text

A system-generated universally unique identifier (UUID) for the action.

budgetName :: Text

notificationType :: NotificationType

actionType :: ActionType

The type of action. This defines the type of tasks that can be carried out by this action. This field also determines the format for definition.

actionThreshold :: ActionThreshold

The trigger threshold of the action.

definition :: Definition

Where you specify all of the type-specific parameters.

executionRoleArn :: Text

The role passed for action execution and reversion. Roles and actions must be in the same account.

approvalModel :: ApprovalModel

This specifies if the action needs manual or automatic approval.

status :: ActionStatus

The status of action.

subscribers :: NonEmpty Subscriber

newAction

Arguments

:: Text
$sel:actionId:Action'

-> Text
$sel:budgetName:Action'

-> NotificationType
$sel:notificationType:Action'

-> ActionType
$sel:actionType:Action'

-> ActionThreshold
$sel:actionThreshold:Action'

-> Definition
$sel:definition:Action'

-> Text
$sel:executionRoleArn:Action'

-> ApprovalModel
$sel:approvalModel:Action'

-> ActionStatus
$sel:status:Action'

-> NonEmpty Subscriber
$sel:subscribers:Action'

-> Action

Create a value of Action with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionId:Action', action_actionId - A system-generated universally unique identifier (UUID) for the action.

$sel:budgetName:Action', action_budgetName - Undocumented member.

$sel:notificationType:Action', action_notificationType - Undocumented member.

$sel:actionType:Action', action_actionType - The type of action. This defines the type of tasks that can be carried out by this action. This field also determines the format for definition.

$sel:actionThreshold:Action', action_actionThreshold - The trigger threshold of the action.

$sel:definition:Action', action_definition - Where you specify all of the type-specific parameters.

$sel:executionRoleArn:Action', action_executionRoleArn - The role passed for action execution and reversion. Roles and actions must be in the same account.

$sel:approvalModel:Action', action_approvalModel - This specifies if the action needs manual or automatic approval.

$sel:status:Action', action_status - The status of action.

$sel:subscribers:Action', action_subscribers - Undocumented member.

action_actionId :: Lens' Action Text

A system-generated universally unique identifier (UUID) for the action.

action_budgetName :: Lens' Action Text

Undocumented member.

action_notificationType :: Lens' Action NotificationType

Undocumented member.

action_actionType :: Lens' Action ActionType

The type of action. This defines the type of tasks that can be carried out by this action. This field also determines the format for definition.

action_actionThreshold :: Lens' Action ActionThreshold

The trigger threshold of the action.

action_definition :: Lens' Action Definition

Where you specify all of the type-specific parameters.

action_executionRoleArn :: Lens' Action Text

The role passed for action execution and reversion. Roles and actions must be in the same account.

action_approvalModel :: Lens' Action ApprovalModel

This specifies if the action needs manual or automatic approval.

action_status :: Lens' Action ActionStatus

The status of action.

action_subscribers :: Lens' Action (NonEmpty Subscriber)

Undocumented member.