Network.AWS.Budgets.Types.ActionHistory

Description

Documentation

data ActionHistory

The historical records for a budget action.

See: newActionHistory smart constructor.

Constructors

ActionHistory'

Fields

timestamp :: POSIX

status :: ActionStatus

The status of action at the time of the event.

eventType :: EventType

This distinguishes between whether the events are triggered by the user or generated by the system.

actionHistoryDetails :: ActionHistoryDetails

The description of details of the event.

newActionHistory

Arguments

:: UTCTime
$sel:timestamp:ActionHistory'

-> ActionStatus
$sel:status:ActionHistory'

-> EventType
$sel:eventType:ActionHistory'

-> ActionHistoryDetails
$sel:actionHistoryDetails:ActionHistory'

-> ActionHistory

Create a value of ActionHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timestamp:ActionHistory', actionHistory_timestamp - Undocumented member.

$sel:status:ActionHistory', actionHistory_status - The status of action at the time of the event.

$sel:eventType:ActionHistory', actionHistory_eventType - This distinguishes between whether the events are triggered by the user or generated by the system.

$sel:actionHistoryDetails:ActionHistory', actionHistory_actionHistoryDetails - The description of details of the event.

actionHistory_timestamp :: Lens' ActionHistory UTCTime

Undocumented member.

actionHistory_status :: Lens' ActionHistory ActionStatus

The status of action at the time of the event.

actionHistory_eventType :: Lens' ActionHistory EventType

This distinguishes between whether the events are triggered by the user or generated by the system.

actionHistory_actionHistoryDetails :: Lens' ActionHistory ActionHistoryDetails

The description of details of the event.