Network.AWS.Budgets.Types.BudgetPerformanceHistory

Description

Documentation

data BudgetPerformanceHistory

A history of the state of a budget at the end of the budget's specified time period.

See: newBudgetPerformanceHistory smart constructor.

Constructors

BudgetPerformanceHistory'

Fields

budgetedAndActualAmountsList :: Maybe [BudgetedAndActualAmounts]

A list of amounts of cost or usage that you created budgets for, compared to your actual costs or usage.

timeUnit :: Maybe TimeUnit

costFilters :: Maybe (HashMap Text [Text])

The history of the cost filters for a budget during the specified time period.

costTypes :: Maybe CostTypes

The history of the cost types for a budget during the specified time period.

budgetType :: Maybe BudgetType

budgetName :: Maybe Text

newBudgetPerformanceHistory :: BudgetPerformanceHistory

Create a value of BudgetPerformanceHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:budgetedAndActualAmountsList:BudgetPerformanceHistory', budgetPerformanceHistory_budgetedAndActualAmountsList - A list of amounts of cost or usage that you created budgets for, compared to your actual costs or usage.

$sel:timeUnit:BudgetPerformanceHistory', budgetPerformanceHistory_timeUnit - Undocumented member.

$sel:costFilters:BudgetPerformanceHistory', budgetPerformanceHistory_costFilters - The history of the cost filters for a budget during the specified time period.

$sel:costTypes:BudgetPerformanceHistory', budgetPerformanceHistory_costTypes - The history of the cost types for a budget during the specified time period.

$sel:budgetType:BudgetPerformanceHistory', budgetPerformanceHistory_budgetType - Undocumented member.

$sel:budgetName:BudgetPerformanceHistory', budgetPerformanceHistory_budgetName - Undocumented member.

budgetPerformanceHistory_budgetedAndActualAmountsList :: Lens' BudgetPerformanceHistory (Maybe [BudgetedAndActualAmounts])

A list of amounts of cost or usage that you created budgets for, compared to your actual costs or usage.

budgetPerformanceHistory_timeUnit :: Lens' BudgetPerformanceHistory (Maybe TimeUnit)

Undocumented member.

budgetPerformanceHistory_costFilters :: Lens' BudgetPerformanceHistory (Maybe (HashMap Text [Text]))

The history of the cost filters for a budget during the specified time period.

budgetPerformanceHistory_costTypes :: Lens' BudgetPerformanceHistory (Maybe CostTypes)

The history of the cost types for a budget during the specified time period.

budgetPerformanceHistory_budgetType :: Lens' BudgetPerformanceHistory (Maybe BudgetType)

Undocumented member.

budgetPerformanceHistory_budgetName :: Lens' BudgetPerformanceHistory (Maybe Text)

Undocumented member.