Network.AWS.Budgets.Types.BudgetedAndActualAmounts

Description

Documentation

data BudgetedAndActualAmounts

The amount of cost or usage that you created the budget for, compared to your actual costs or usage.

See: newBudgetedAndActualAmounts smart constructor.

Constructors

BudgetedAndActualAmounts'

Fields

timePeriod :: Maybe TimePeriod

The time period covered by this budget comparison.

budgetedAmount :: Maybe Spend

The amount of cost or usage that you created the budget for.

actualAmount :: Maybe Spend

Your actual costs or usage for a budget period.

newBudgetedAndActualAmounts :: BudgetedAndActualAmounts

Create a value of BudgetedAndActualAmounts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timePeriod:BudgetedAndActualAmounts', budgetedAndActualAmounts_timePeriod - The time period covered by this budget comparison.

$sel:budgetedAmount:BudgetedAndActualAmounts', budgetedAndActualAmounts_budgetedAmount - The amount of cost or usage that you created the budget for.

$sel:actualAmount:BudgetedAndActualAmounts', budgetedAndActualAmounts_actualAmount - Your actual costs or usage for a budget period.

budgetedAndActualAmounts_timePeriod :: Lens' BudgetedAndActualAmounts (Maybe TimePeriod)

The time period covered by this budget comparison.

budgetedAndActualAmounts_budgetedAmount :: Lens' BudgetedAndActualAmounts (Maybe Spend)

The amount of cost or usage that you created the budget for.

budgetedAndActualAmounts_actualAmount :: Lens' BudgetedAndActualAmounts (Maybe Spend)

Your actual costs or usage for a budget period.