Network.AWS.Budgets.Types.CostTypes

Description

Documentation

data CostTypes

The types of cost that are included in a COST budget, such as tax and subscriptions.

USAGE, RI_UTILIZATION, RI_COVERAGE, SAVINGS_PLANS_UTILIZATION, and SAVINGS_PLANS_COVERAGE budgets do not have CostTypes.

See: newCostTypes smart constructor.

Constructors

CostTypes'

Fields

useAmortized :: Maybe Bool

Specifies whether a budget uses the amortized rate.

The default value is false.

includeRecurring :: Maybe Bool

Specifies whether a budget includes recurring fees such as monthly RI fees.

The default value is true.

useBlended :: Maybe Bool

Specifies whether a budget uses a blended rate.

The default value is false.

includeSupport :: Maybe Bool

Specifies whether a budget includes support subscription fees.

The default value is true.

includeDiscount :: Maybe Bool

Specifies whether a budget includes discounts.

The default value is true.

includeSubscription :: Maybe Bool

Specifies whether a budget includes subscriptions.

The default value is true.

includeRefund :: Maybe Bool

Specifies whether a budget includes refunds.

The default value is true.

includeUpfront :: Maybe Bool

Specifies whether a budget includes upfront RI costs.

The default value is true.

includeOtherSubscription :: Maybe Bool

Specifies whether a budget includes non-RI subscription costs.

The default value is true.

includeTax :: Maybe Bool

Specifies whether a budget includes taxes.

The default value is true.

includeCredit :: Maybe Bool

Specifies whether a budget includes credits.

The default value is true.

newCostTypes :: CostTypes

Create a value of CostTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:useAmortized:CostTypes', costTypes_useAmortized - Specifies whether a budget uses the amortized rate.

The default value is false.

$sel:includeRecurring:CostTypes', costTypes_includeRecurring - Specifies whether a budget includes recurring fees such as monthly RI fees.

The default value is true.

$sel:useBlended:CostTypes', costTypes_useBlended - Specifies whether a budget uses a blended rate.

The default value is false.

$sel:includeSupport:CostTypes', costTypes_includeSupport - Specifies whether a budget includes support subscription fees.

The default value is true.

$sel:includeDiscount:CostTypes', costTypes_includeDiscount - Specifies whether a budget includes discounts.

The default value is true.

$sel:includeSubscription:CostTypes', costTypes_includeSubscription - Specifies whether a budget includes subscriptions.

The default value is true.

$sel:includeRefund:CostTypes', costTypes_includeRefund - Specifies whether a budget includes refunds.

The default value is true.

$sel:includeUpfront:CostTypes', costTypes_includeUpfront - Specifies whether a budget includes upfront RI costs.

The default value is true.

$sel:includeOtherSubscription:CostTypes', costTypes_includeOtherSubscription - Specifies whether a budget includes non-RI subscription costs.

The default value is true.

$sel:includeTax:CostTypes', costTypes_includeTax - Specifies whether a budget includes taxes.

The default value is true.

$sel:includeCredit:CostTypes', costTypes_includeCredit - Specifies whether a budget includes credits.

The default value is true.

costTypes_useAmortized :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget uses the amortized rate.

The default value is false.

costTypes_includeRecurring :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes recurring fees such as monthly RI fees.

The default value is true.

costTypes_useBlended :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget uses a blended rate.

The default value is false.

costTypes_includeSupport :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes support subscription fees.

The default value is true.

costTypes_includeDiscount :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes discounts.

The default value is true.

costTypes_includeSubscription :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes subscriptions.

The default value is true.

costTypes_includeRefund :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes refunds.

The default value is true.

costTypes_includeUpfront :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes upfront RI costs.

The default value is true.

costTypes_includeOtherSubscription :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes non-RI subscription costs.

The default value is true.

costTypes_includeTax :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes taxes.

The default value is true.

costTypes_includeCredit :: Lens' CostTypes (Maybe Bool)

Specifies whether a budget includes credits.

The default value is true.