Network.AWS.Budgets.Types.Spend

Description

Documentation

data Spend

The amount of cost or usage that is measured for a budget.

For example, a Spend for 3 GB of S3 usage would have the following parameters:

An Amount of 3

A unit of GB

See: newSpend smart constructor.

Constructors

Spend'

Fields

amount :: Text

The cost or usage amount that is associated with a budget forecast, actual spend, or budget threshold.

unit :: Text

The unit of measurement that is used for the budget forecast, actual spend, or budget threshold, such as dollars or GB.

newSpend

Arguments

:: Text
$sel:amount:Spend'

-> Text
$sel:unit:Spend'

-> Spend

Create a value of Spend with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amount:Spend', spend_amount - The cost or usage amount that is associated with a budget forecast, actual spend, or budget threshold.

$sel:unit:Spend', spend_unit - The unit of measurement that is used for the budget forecast, actual spend, or budget threshold, such as dollars or GB.

spend_amount :: Lens' Spend Text

The cost or usage amount that is associated with a budget forecast, actual spend, or budget threshold.

spend_unit :: Lens' Spend Text

The unit of measurement that is used for the budget forecast, actual spend, or budget threshold, such as dollars or GB.