Network.AWS.Budgets.UpdateBudgetAction

Description

Updates a budget action.

Creating a Request

data UpdateBudgetAction

See: newUpdateBudgetAction smart constructor.

Constructors

UpdateBudgetAction'

Fields

subscribers :: Maybe (NonEmpty Subscriber)

executionRoleArn :: Maybe Text

The role passed for action execution and reversion. Roles and actions must be in the same account.

approvalModel :: Maybe ApprovalModel

This specifies if the action needs manual or automatic approval.

notificationType :: Maybe NotificationType

actionThreshold :: Maybe ActionThreshold

definition :: Maybe Definition

accountId :: Text

budgetName :: Text

actionId :: Text

A system-generated universally unique identifier (UUID) for the action.

newUpdateBudgetAction

Arguments

:: Text
$sel:accountId:UpdateBudgetAction'

-> Text
$sel:budgetName:UpdateBudgetAction'

-> Text
$sel:actionId:UpdateBudgetAction'

-> UpdateBudgetAction

Create a value of UpdateBudgetAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscribers:UpdateBudgetAction', updateBudgetAction_subscribers - Undocumented member.

$sel:executionRoleArn:UpdateBudgetAction', updateBudgetAction_executionRoleArn - The role passed for action execution and reversion. Roles and actions must be in the same account.

$sel:approvalModel:UpdateBudgetAction', updateBudgetAction_approvalModel - This specifies if the action needs manual or automatic approval.

$sel:notificationType:UpdateBudgetAction', updateBudgetAction_notificationType - Undocumented member.

$sel:actionThreshold:UpdateBudgetAction', updateBudgetAction_actionThreshold - Undocumented member.

$sel:definition:UpdateBudgetAction', updateBudgetAction_definition - Undocumented member.

$sel:accountId:UpdateBudgetAction', updateBudgetAction_accountId - Undocumented member.

$sel:budgetName:UpdateBudgetAction', updateBudgetAction_budgetName - Undocumented member.

$sel:actionId:UpdateBudgetAction', updateBudgetAction_actionId - A system-generated universally unique identifier (UUID) for the action.

Request Lenses

updateBudgetAction_subscribers :: Lens' UpdateBudgetAction (Maybe (NonEmpty Subscriber))

Undocumented member.

updateBudgetAction_executionRoleArn :: Lens' UpdateBudgetAction (Maybe Text)

The role passed for action execution and reversion. Roles and actions must be in the same account.

updateBudgetAction_approvalModel :: Lens' UpdateBudgetAction (Maybe ApprovalModel)

This specifies if the action needs manual or automatic approval.

updateBudgetAction_notificationType :: Lens' UpdateBudgetAction (Maybe NotificationType)

Undocumented member.

updateBudgetAction_actionThreshold :: Lens' UpdateBudgetAction (Maybe ActionThreshold)

Undocumented member.

updateBudgetAction_definition :: Lens' UpdateBudgetAction (Maybe Definition)

Undocumented member.

updateBudgetAction_accountId :: Lens' UpdateBudgetAction Text

Undocumented member.

updateBudgetAction_budgetName :: Lens' UpdateBudgetAction Text

Undocumented member.

updateBudgetAction_actionId :: Lens' UpdateBudgetAction Text

A system-generated universally unique identifier (UUID) for the action.

Destructuring the Response

data UpdateBudgetActionResponse

See: newUpdateBudgetActionResponse smart constructor.

Constructors

UpdateBudgetActionResponse'

Fields

httpStatus :: Int

The response's http status code.

accountId :: Text

budgetName :: Text

oldAction :: Action

The previous action resource information.

newAction' :: Action

The updated action resource information.

newUpdateBudgetActionResponse

Arguments

:: Int
$sel:httpStatus:UpdateBudgetActionResponse'

-> Text
$sel:accountId:UpdateBudgetAction'

-> Text
$sel:budgetName:UpdateBudgetAction'

-> Action
$sel:oldAction:UpdateBudgetActionResponse'

-> Action
$sel:newAction':UpdateBudgetActionResponse'

-> UpdateBudgetActionResponse

Create a value of UpdateBudgetActionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateBudgetActionResponse', updateBudgetActionResponse_httpStatus - The response's http status code.

$sel:accountId:UpdateBudgetAction', updateBudgetActionResponse_accountId - Undocumented member.

$sel:budgetName:UpdateBudgetAction', updateBudgetActionResponse_budgetName - Undocumented member.

$sel:oldAction:UpdateBudgetActionResponse', updateBudgetActionResponse_oldAction - The previous action resource information.

$sel:newAction':UpdateBudgetActionResponse', updateBudgetActionResponse_newAction - The updated action resource information.

Response Lenses

updateBudgetActionResponse_httpStatus :: Lens' UpdateBudgetActionResponse Int

The response's http status code.

updateBudgetActionResponse_accountId :: Lens' UpdateBudgetActionResponse Text

Undocumented member.

updateBudgetActionResponse_budgetName :: Lens' UpdateBudgetActionResponse Text

Undocumented member.

updateBudgetActionResponse_oldAction :: Lens' UpdateBudgetActionResponse Action

The previous action resource information.

updateBudgetActionResponse_newAction :: Lens' UpdateBudgetActionResponse Action

The updated action resource information.