Network.AWS.Budgets.UpdateSubscriber

Description

Updates a subscriber.

Creating a Request

data UpdateSubscriber

Request of UpdateSubscriber

See: newUpdateSubscriber smart constructor.

Constructors

UpdateSubscriber'

Fields

accountId :: Text

The accountId that is associated with the budget whose subscriber you want to update.

budgetName :: Text

The name of the budget whose subscriber you want to update.

notification :: Notification

The notification whose subscriber you want to update.

oldSubscriber :: Subscriber

The previous subscriber that is associated with a budget notification.

newSubscriber' :: Subscriber

The updated subscriber that is associated with a budget notification.

newUpdateSubscriber

Arguments

:: Text
$sel:accountId:UpdateSubscriber'

-> Text
$sel:budgetName:UpdateSubscriber'

-> Notification
$sel:notification:UpdateSubscriber'

-> Subscriber
$sel:oldSubscriber:UpdateSubscriber'

-> Subscriber
$sel:newSubscriber':UpdateSubscriber'

-> UpdateSubscriber

Create a value of UpdateSubscriber with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:UpdateSubscriber', updateSubscriber_accountId - The accountId that is associated with the budget whose subscriber you want to update.

$sel:budgetName:UpdateSubscriber', updateSubscriber_budgetName - The name of the budget whose subscriber you want to update.

$sel:notification:UpdateSubscriber', updateSubscriber_notification - The notification whose subscriber you want to update.

$sel:oldSubscriber:UpdateSubscriber', updateSubscriber_oldSubscriber - The previous subscriber that is associated with a budget notification.

$sel:newSubscriber':UpdateSubscriber', updateSubscriber_newSubscriber - The updated subscriber that is associated with a budget notification.

Request Lenses

updateSubscriber_accountId :: Lens' UpdateSubscriber Text

The accountId that is associated with the budget whose subscriber you want to update.

updateSubscriber_budgetName :: Lens' UpdateSubscriber Text

The name of the budget whose subscriber you want to update.

updateSubscriber_notification :: Lens' UpdateSubscriber Notification

The notification whose subscriber you want to update.

updateSubscriber_oldSubscriber :: Lens' UpdateSubscriber Subscriber

The previous subscriber that is associated with a budget notification.

updateSubscriber_newSubscriber :: Lens' UpdateSubscriber Subscriber

The updated subscriber that is associated with a budget notification.

Destructuring the Response

data UpdateSubscriberResponse

Response of UpdateSubscriber

See: newUpdateSubscriberResponse smart constructor.

Constructors

UpdateSubscriberResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateSubscriberResponse

Arguments

:: Int
$sel:httpStatus:UpdateSubscriberResponse'

-> UpdateSubscriberResponse

Create a value of UpdateSubscriberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateSubscriberResponse', updateSubscriberResponse_httpStatus - The response's http status code.

Response Lenses

updateSubscriberResponse_httpStatus :: Lens' UpdateSubscriberResponse Int

The response's http status code.