Network.AWS.CertificateManager.GetAccountConfiguration

Description

Returns the account configuration options associated with an Amazon Web Services account.

Creating a Request

data GetAccountConfiguration

See: newGetAccountConfiguration smart constructor.

Constructors

GetAccountConfiguration'

newGetAccountConfiguration :: GetAccountConfiguration

Create a value of GetAccountConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetAccountConfigurationResponse

See: newGetAccountConfigurationResponse smart constructor.

Constructors

GetAccountConfigurationResponse'

Fields

expiryEvents :: Maybe ExpiryEventsConfiguration

Expiration events configuration options associated with the Amazon Web Services account.

httpStatus :: Int

The response's http status code.

newGetAccountConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetAccountConfigurationResponse'

-> GetAccountConfigurationResponse

Create a value of GetAccountConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expiryEvents:GetAccountConfigurationResponse', getAccountConfigurationResponse_expiryEvents - Expiration events configuration options associated with the Amazon Web Services account.

$sel:httpStatus:GetAccountConfigurationResponse', getAccountConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getAccountConfigurationResponse_expiryEvents :: Lens' GetAccountConfigurationResponse (Maybe ExpiryEventsConfiguration)

Expiration events configuration options associated with the Amazon Web Services account.

getAccountConfigurationResponse_httpStatus :: Lens' GetAccountConfigurationResponse Int

The response's http status code.