Network.AWS.CertificateManager.Lens

Description

Operations

GetAccountConfiguration

getAccountConfigurationResponse_expiryEvents :: Lens' GetAccountConfigurationResponse (Maybe ExpiryEventsConfiguration)

Expiration events configuration options associated with the Amazon Web Services account.

getAccountConfigurationResponse_httpStatus :: Lens' GetAccountConfigurationResponse Int

The response's http status code.

RemoveTagsFromCertificate

removeTagsFromCertificate_certificateArn :: Lens' RemoveTagsFromCertificate Text

String that contains the ARN of the ACM Certificate with one or more tags that you want to remove. This must be of the form:

arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012

For more information about ARNs, see Amazon Resource Names (ARNs).

removeTagsFromCertificate_tags :: Lens' RemoveTagsFromCertificate (NonEmpty Tag)

The key-value pair that defines the tag to remove.

DeleteCertificate

deleteCertificate_certificateArn :: Lens' DeleteCertificate Text

String that contains the ARN of the ACM certificate to be deleted. This must be of the form:

arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012

For more information about ARNs, see Amazon Resource Names (ARNs).

UpdateCertificateOptions

updateCertificateOptions_certificateArn :: Lens' UpdateCertificateOptions Text

ARN of the requested certificate to update. This must be of the form:

arn:aws:acm:us-east-1:account:certificate/12345678-1234-1234-1234-123456789012

updateCertificateOptions_options :: Lens' UpdateCertificateOptions CertificateOptions

Use to update the options for your certificate. Currently, you can specify whether to add your certificate to a transparency log. Certificate transparency makes it possible to detect SSL/TLS certificates that have been mistakenly or maliciously issued. Certificates that have not been logged typically produce an error message in a browser.

ExportCertificate

exportCertificate_certificateArn :: Lens' ExportCertificate Text

An Amazon Resource Name (ARN) of the issued certificate. This must be of the form:

arn:aws:acm:region:account:certificate/12345678-1234-1234-1234-123456789012

exportCertificate_passphrase :: Lens' ExportCertificate ByteString

Passphrase to associate with the encrypted exported private key. If you want to later decrypt the private key, you must have the passphrase. You can use the following OpenSSL command to decrypt a private key:

openssl rsa -in encrypted_key.pem -out decrypted_key.pem-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

exportCertificateResponse_privateKey :: Lens' ExportCertificateResponse (Maybe Text)

The encrypted private key associated with the public key in the certificate. The key is output in PKCS #8 format and is base64 PEM-encoded.

exportCertificateResponse_certificateChain :: Lens' ExportCertificateResponse (Maybe Text)

The base64 PEM-encoded certificate chain. This does not include the certificate that you are exporting.

exportCertificateResponse_certificate :: Lens' ExportCertificateResponse (Maybe Text)

The base64 PEM-encoded certificate.

exportCertificateResponse_httpStatus :: Lens' ExportCertificateResponse Int

The response's http status code.

RenewCertificate

renewCertificate_certificateArn :: Lens' RenewCertificate Text

String that contains the ARN of the ACM certificate to be renewed. This must be of the form:

arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012

For more information about ARNs, see Amazon Resource Names (ARNs).

GetCertificate

getCertificate_certificateArn :: Lens' GetCertificate Text

String that contains a certificate ARN in the following format:

arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012

For more information about ARNs, see Amazon Resource Names (ARNs).

getCertificateResponse_certificateChain :: Lens' GetCertificateResponse (Maybe Text)

Certificates forming the requested certificate's chain of trust. The chain consists of the certificate of the issuing CA and the intermediate certificates of any other subordinate CAs.

getCertificateResponse_certificate :: Lens' GetCertificateResponse (Maybe Text)

The ACM-issued certificate corresponding to the ARN specified as input.

getCertificateResponse_httpStatus :: Lens' GetCertificateResponse Int

The response's http status code.

ImportCertificate

importCertificate_certificateArn :: Lens' ImportCertificate (Maybe Text)

The Amazon Resource Name (ARN) of an imported certificate to replace. To import a new certificate, omit this field.

importCertificate_tags :: Lens' ImportCertificate (Maybe (NonEmpty Tag))

One or more resource tags to associate with the imported certificate.

Note: You cannot apply tags when reimporting a certificate.

importCertificate_certificateChain :: Lens' ImportCertificate (Maybe ByteString)

The PEM encoded certificate chain.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

importCertificate_certificate :: Lens' ImportCertificate ByteString

The certificate to import.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

importCertificate_privateKey :: Lens' ImportCertificate ByteString

The private key that matches the public key in the certificate.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

importCertificateResponse_certificateArn :: Lens' ImportCertificateResponse (Maybe Text)

The Amazon Resource Name (ARN) of the imported certificate.

importCertificateResponse_httpStatus :: Lens' ImportCertificateResponse Int

The response's http status code.

DescribeCertificate

describeCertificate_certificateArn :: Lens' DescribeCertificate Text

The Amazon Resource Name (ARN) of the ACM certificate. The ARN must have the following form:

arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012

For more information about ARNs, see Amazon Resource Names (ARNs).

describeCertificateResponse_certificate :: Lens' DescribeCertificateResponse (Maybe CertificateDetail)

Metadata about an ACM certificate.

describeCertificateResponse_httpStatus :: Lens' DescribeCertificateResponse Int

The response's http status code.

PutAccountConfiguration

putAccountConfiguration_expiryEvents :: Lens' PutAccountConfiguration (Maybe ExpiryEventsConfiguration)

Specifies expiration events associated with an account.

putAccountConfiguration_idempotencyToken :: Lens' PutAccountConfiguration Text

Customer-chosen string used to distinguish between calls to PutAccountConfiguration. Idempotency tokens time out after one hour. If you call PutAccountConfiguration multiple times with the same unexpired idempotency token, ACM treats it as the same request and returns the original result. If you change the idempotency token for each call, ACM treats each call as a new request.

ResendValidationEmail

resendValidationEmail_certificateArn :: Lens' ResendValidationEmail Text

String that contains the ARN of the requested certificate. The certificate ARN is generated and returned by the RequestCertificate action as soon as the request is made. By default, using this parameter causes email to be sent to all top-level domains you specified in the certificate request. The ARN must be of the form:

arn:aws:acm:us-east-1:123456789012:certificate/12345678-1234-1234-1234-123456789012

resendValidationEmail_domain :: Lens' ResendValidationEmail Text

The fully qualified domain name (FQDN) of the certificate that needs to be validated.

resendValidationEmail_validationDomain :: Lens' ResendValidationEmail Text

The base validation domain that will act as the suffix of the email addresses that are used to send the emails. This must be the same as the Domain value or a superdomain of the Domain value. For example, if you requested a certificate for site.subdomain.example.com and specify a ValidationDomain of subdomain.example.com, ACM sends email to the domain registrant, technical contact, and administrative contact in WHOIS and the following five addresses:

admin@subdomain.example.com

administrator@subdomain.example.com

hostmaster@subdomain.example.com

postmaster@subdomain.example.com

webmaster@subdomain.example.com

ListCertificates

listCertificates_nextToken :: Lens' ListCertificates (Maybe Text)

Use this parameter only when paginating results and only in a subsequent request after you receive a response with truncated results. Set it to the value of NextToken from the response you just received.

listCertificates_certificateStatuses :: Lens' ListCertificates (Maybe [CertificateStatus])

Filter the certificate list by status value.

listCertificates_includes :: Lens' ListCertificates (Maybe Filters)

Filter the certificate list. For more information, see the Filters structure.

listCertificates_maxItems :: Lens' ListCertificates (Maybe Natural)

Use this parameter when paginating results to specify the maximum number of items to return in the response. If additional items exist beyond the number you specify, the NextToken element is sent in the response. Use this NextToken value in a subsequent request to retrieve additional items.

listCertificatesResponse_nextToken :: Lens' ListCertificatesResponse (Maybe Text)

When the list is truncated, this value is present and contains the value to use for the NextToken parameter in a subsequent pagination request.

listCertificatesResponse_certificateSummaryList :: Lens' ListCertificatesResponse (Maybe [CertificateSummary])

A list of ACM certificates.

listCertificatesResponse_httpStatus :: Lens' ListCertificatesResponse Int

The response's http status code.

RequestCertificate

requestCertificate_idempotencyToken :: Lens' RequestCertificate (Maybe Text)

Customer chosen string that can be used to distinguish between calls to RequestCertificate. Idempotency tokens time out after one hour. Therefore, if you call RequestCertificate multiple times with the same idempotency token within one hour, ACM recognizes that you are requesting only one certificate and will issue only one. If you change the idempotency token for each call, ACM recognizes that you are requesting multiple certificates.

requestCertificate_validationMethod :: Lens' RequestCertificate (Maybe ValidationMethod)

The method you want to use if you are requesting a public certificate to validate that you own or control domain. You can validate with DNS or validate with email. We recommend that you use DNS validation.

requestCertificate_certificateAuthorityArn :: Lens' RequestCertificate (Maybe Text)

The Amazon Resource Name (ARN) of the private certificate authority (CA) that will be used to issue the certificate. If you do not provide an ARN and you are trying to request a private certificate, ACM will attempt to issue a public certificate. For more information about private CAs, see the Amazon Web Services Certificate Manager Private Certificate Authority (PCA) user guide. The ARN must have the following form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

requestCertificate_domainValidationOptions :: Lens' RequestCertificate (Maybe (NonEmpty DomainValidationOption))

The domain name that you want ACM to use to send you emails so that you can validate domain ownership.

requestCertificate_options :: Lens' RequestCertificate (Maybe CertificateOptions)

Currently, you can use this parameter to specify whether to add the certificate to a certificate transparency log. Certificate transparency makes it possible to detect SSL/TLS certificates that have been mistakenly or maliciously issued. Certificates that have not been logged typically produce an error message in a browser. For more information, see Opting Out of Certificate Transparency Logging.

requestCertificate_subjectAlternativeNames :: Lens' RequestCertificate (Maybe (NonEmpty Text))

Additional FQDNs to be included in the Subject Alternative Name extension of the ACM certificate. For example, add the name www.example.net to a certificate for which the DomainName field is www.example.com if users can reach your site by using either name. The maximum number of domain names that you can add to an ACM certificate is 100. However, the initial quota is 10 domain names. If you need more than 10 names, you must request a quota increase. For more information, see Quotas.

The maximum length of a SAN DNS name is 253 octets. The name is made up of multiple labels separated by periods. No label can be longer than 63 octets. Consider the following examples:

(63 octets).(63 octets).(63 octets).(61 octets) is legal because the total length is 253 octets (63+1+63+1+63+1+61) and no label exceeds 63 octets.

(64 octets).(63 octets).(63 octets).(61 octets) is not legal because the total length exceeds 253 octets (64+1+63+1+63+1+61) and the first label exceeds 63 octets.

(63 octets).(63 octets).(63 octets).(62 octets) is not legal because the total length of the DNS name (63+1+63+1+63+1+62) exceeds 253 octets.

requestCertificate_tags :: Lens' RequestCertificate (Maybe (NonEmpty Tag))

One or more resource tags to associate with the certificate.

requestCertificate_domainName :: Lens' RequestCertificate Text

Fully qualified domain name (FQDN), such as www.example.com, that you want to secure with an ACM certificate. Use an asterisk (*) to create a wildcard certificate that protects several sites in the same domain. For example, *.example.com protects www.example.com, site.example.com, and images.example.com.

The first domain name you enter cannot exceed 64 octets, including periods. Each subsequent Subject Alternative Name (SAN), however, can be up to 253 octets in length.

requestCertificateResponse_certificateArn :: Lens' RequestCertificateResponse (Maybe Text)

String that contains the ARN of the issued certificate. This must be of the form:

arn:aws:acm:us-east-1:123456789012:certificate/12345678-1234-1234-1234-123456789012

requestCertificateResponse_httpStatus :: Lens' RequestCertificateResponse Int

The response's http status code.

AddTagsToCertificate

addTagsToCertificate_certificateArn :: Lens' AddTagsToCertificate Text

String that contains the ARN of the ACM certificate to which the tag is to be applied. This must be of the form:

arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012

For more information about ARNs, see Amazon Resource Names (ARNs).

addTagsToCertificate_tags :: Lens' AddTagsToCertificate (NonEmpty Tag)

The key-value pair that defines the tag. The tag value is optional.

ListTagsForCertificate

listTagsForCertificate_certificateArn :: Lens' ListTagsForCertificate Text

String that contains the ARN of the ACM certificate for which you want to list the tags. This must have the following form:

arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012

For more information about ARNs, see Amazon Resource Names (ARNs).

listTagsForCertificateResponse_tags :: Lens' ListTagsForCertificateResponse (Maybe (NonEmpty Tag))

The key-value pairs that define the applied tags.

listTagsForCertificateResponse_httpStatus :: Lens' ListTagsForCertificateResponse Int

The response's http status code.

Types

CertificateDetail

certificateDetail_status :: Lens' CertificateDetail (Maybe CertificateStatus)

The status of the certificate.

certificateDetail_notBefore :: Lens' CertificateDetail (Maybe UTCTime)

The time before which the certificate is not valid.

certificateDetail_certificateAuthorityArn :: Lens' CertificateDetail (Maybe Text)

The Amazon Resource Name (ARN) of the ACM PCA private certificate authority (CA) that issued the certificate. This has the following format:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

certificateDetail_importedAt :: Lens' CertificateDetail (Maybe UTCTime)

The date and time at which the certificate was imported. This value exists only when the certificate type is IMPORTED.

certificateDetail_extendedKeyUsages :: Lens' CertificateDetail (Maybe [ExtendedKeyUsage])

Contains a list of Extended Key Usage X.509 v3 extension objects. Each object specifies a purpose for which the certificate public key can be used and consists of a name and an object identifier (OID).

certificateDetail_domainValidationOptions :: Lens' CertificateDetail (Maybe (NonEmpty DomainValidation))

Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED.

certificateDetail_renewalEligibility :: Lens' CertificateDetail (Maybe RenewalEligibility)

Specifies whether the certificate is eligible for renewal. At this time, only exported private certificates can be renewed with the RenewCertificate command.

certificateDetail_options :: Lens' CertificateDetail (Maybe CertificateOptions)

Value that specifies whether to add the certificate to a transparency log. Certificate transparency makes it possible to detect SSL certificates that have been mistakenly or maliciously issued. A browser might respond to certificate that has not been logged by showing an error message. The logs are cryptographically secure.

certificateDetail_serial :: Lens' CertificateDetail (Maybe Text)

The serial number of the certificate.

certificateDetail_certificateArn :: Lens' CertificateDetail (Maybe Text)

The Amazon Resource Name (ARN) of the certificate. For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

certificateDetail_createdAt :: Lens' CertificateDetail (Maybe UTCTime)

The time at which the certificate was requested.

certificateDetail_inUseBy :: Lens' CertificateDetail (Maybe [Text])

A list of ARNs for the Amazon Web Services resources that are using the certificate. A certificate can be used by multiple Amazon Web Services resources.

certificateDetail_domainName :: Lens' CertificateDetail (Maybe Text)

The fully qualified domain name for the certificate, such as www.example.com or example.com.

certificateDetail_subjectAlternativeNames :: Lens' CertificateDetail (Maybe (NonEmpty Text))

One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate. The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

certificateDetail_revocationReason :: Lens' CertificateDetail (Maybe RevocationReason)

The reason the certificate was revoked. This value exists only when the certificate status is REVOKED.

certificateDetail_failureReason :: Lens' CertificateDetail (Maybe FailureReason)

The reason the certificate request failed. This value exists only when the certificate status is FAILED. For more information, see Certificate Request Failed in the Amazon Web Services Certificate Manager User Guide.

certificateDetail_subject :: Lens' CertificateDetail (Maybe Text)

The name of the entity that is associated with the public key contained in the certificate.

certificateDetail_notAfter :: Lens' CertificateDetail (Maybe UTCTime)

The time after which the certificate is not valid.

certificateDetail_revokedAt :: Lens' CertificateDetail (Maybe UTCTime)

The time at which the certificate was revoked. This value exists only when the certificate status is REVOKED.

certificateDetail_keyUsages :: Lens' CertificateDetail (Maybe [KeyUsage])

A list of Key Usage X.509 v3 extension objects. Each object is a string value that identifies the purpose of the public key contained in the certificate. Possible extension values include DIGITAL_SIGNATURE, KEY_ENCHIPHERMENT, NON_REPUDIATION, and more.

certificateDetail_issuer :: Lens' CertificateDetail (Maybe Text)

The name of the certificate authority that issued and signed the certificate.

certificateDetail_signatureAlgorithm :: Lens' CertificateDetail (Maybe Text)

The algorithm that was used to sign the certificate.

certificateDetail_issuedAt :: Lens' CertificateDetail (Maybe UTCTime)

The time at which the certificate was issued. This value exists only when the certificate type is AMAZON_ISSUED.

certificateDetail_keyAlgorithm :: Lens' CertificateDetail (Maybe KeyAlgorithm)

The algorithm that was used to generate the public-private key pair.

certificateDetail_type :: Lens' CertificateDetail (Maybe CertificateType)

The source of the certificate. For certificates provided by ACM, this value is AMAZON_ISSUED. For certificates that you imported with ImportCertificate, this value is IMPORTED. ACM does not provide managed renewal for imported certificates. For more information about the differences between certificates that you import and those that ACM provides, see Importing Certificates in the Amazon Web Services Certificate Manager User Guide.

certificateDetail_renewalSummary :: Lens' CertificateDetail (Maybe RenewalSummary)

Contains information about the status of ACM's managed renewal for the certificate. This field exists only when the certificate type is AMAZON_ISSUED.

CertificateOptions

certificateOptions_certificateTransparencyLoggingPreference :: Lens' CertificateOptions (Maybe CertificateTransparencyLoggingPreference)

You can opt out of certificate transparency logging by specifying the DISABLED option. Opt in by specifying ENABLED.

CertificateSummary

certificateSummary_certificateArn :: Lens' CertificateSummary (Maybe Text)

Amazon Resource Name (ARN) of the certificate. This is of the form:

arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012

For more information about ARNs, see Amazon Resource Names (ARNs).

certificateSummary_domainName :: Lens' CertificateSummary (Maybe Text)

Fully qualified domain name (FQDN), such as www.example.com or example.com, for the certificate.

DomainValidation

domainValidation_resourceRecord :: Lens' DomainValidation (Maybe ResourceRecord)

Contains the CNAME record that you add to your DNS database for domain validation. For more information, see Use DNS to Validate Domain Ownership.

Note: The CNAME information that you need does not include the name of your domain. If you include your domain name in the DNS database CNAME record, validation fails. For example, if the name is "_a79865eb4cd1a6ab990a45779b4e0b96.yourdomain.com", only "_a79865eb4cd1a6ab990a45779b4e0b96" must be used.

domainValidation_validationMethod :: Lens' DomainValidation (Maybe ValidationMethod)

Specifies the domain validation method.

domainValidation_validationEmails :: Lens' DomainValidation (Maybe [Text])

A list of email addresses that ACM used to send domain validation emails.

domainValidation_validationStatus :: Lens' DomainValidation (Maybe DomainStatus)

The validation status of the domain name. This can be one of the following values:

PENDING_VALIDATION

@@SUCCESS

@@FAILED

domainValidation_validationDomain :: Lens' DomainValidation (Maybe Text)

The domain name that ACM used to send domain validation emails.

domainValidation_domainName :: Lens' DomainValidation Text

A fully qualified domain name (FQDN) in the certificate. For example, www.example.com or example.com.

DomainValidationOption

domainValidationOption_domainName :: Lens' DomainValidationOption Text

A fully qualified domain name (FQDN) in the certificate request.

domainValidationOption_validationDomain :: Lens' DomainValidationOption Text

The domain name that you want ACM to use to send you validation emails. This domain name is the suffix of the email addresses that you want ACM to use. This must be the same as the DomainName value or a superdomain of the DomainName value. For example, if you request a certificate for testing.example.com, you can specify example.com for this value. In that case, ACM sends domain validation emails to the following five addresses:

admin@example.com

administrator@example.com

hostmaster@example.com

postmaster@example.com

webmaster@example.com

ExpiryEventsConfiguration

expiryEventsConfiguration_daysBeforeExpiry :: Lens' ExpiryEventsConfiguration (Maybe Natural)

Specifies the number of days prior to certificate expiration when ACM starts generating EventBridge events. ACM sends one event per day per certificate until the certificate expires. By default, accounts receive events starting 45 days before certificate expiration.

ExtendedKeyUsage

extendedKeyUsage_name :: Lens' ExtendedKeyUsage (Maybe ExtendedKeyUsageName)

The name of an Extended Key Usage value.

extendedKeyUsage_oid :: Lens' ExtendedKeyUsage (Maybe Text)

An object identifier (OID) for the extension value. OIDs are strings of numbers separated by periods. The following OIDs are defined in RFC 3280 and RFC 5280.

1.3.6.1.5.5.7.3.1 (TLS_WEB_SERVER_AUTHENTICATION)

1.3.6.1.5.5.7.3.2 (TLS_WEB_CLIENT_AUTHENTICATION)

1.3.6.1.5.5.7.3.3 (CODE_SIGNING)

1.3.6.1.5.5.7.3.4 (EMAIL_PROTECTION)

1.3.6.1.5.5.7.3.8 (TIME_STAMPING)

1.3.6.1.5.5.7.3.9 (OCSP_SIGNING)

1.3.6.1.5.5.7.3.5 (IPSEC_END_SYSTEM)

1.3.6.1.5.5.7.3.6 (IPSEC_TUNNEL)

1.3.6.1.5.5.7.3.7 (IPSEC_USER)

Filters

filters_keyTypes :: Lens' Filters (Maybe [KeyAlgorithm])

Specify one or more algorithms that can be used to generate key pairs.

Default filtering returns only RSA_1024 and RSA_2048 certificates that have at least one domain. To return other certificate types, provide the desired type signatures in a comma-separated list. For example, "keyTypes": ["RSA_2048,RSA_4096"] returns both RSA_2048 and RSA_4096 certificates.

filters_extendedKeyUsage :: Lens' Filters (Maybe [ExtendedKeyUsageName])

Specify one or more ExtendedKeyUsage extension values.

filters_keyUsage :: Lens' Filters (Maybe [KeyUsageName])

Specify one or more KeyUsage extension values.

KeyUsage

keyUsage_name :: Lens' KeyUsage (Maybe KeyUsageName)

A string value that contains a Key Usage extension name.

RenewalSummary

renewalSummary_renewalStatusReason :: Lens' RenewalSummary (Maybe FailureReason)

The reason that a renewal request was unsuccessful.

renewalSummary_renewalStatus :: Lens' RenewalSummary RenewalStatus

The status of ACM's managed renewal of the certificate.

renewalSummary_domainValidationOptions :: Lens' RenewalSummary (NonEmpty DomainValidation)

Contains information about the validation of each domain name in the certificate, as it pertains to ACM's managed renewal. This is different from the initial validation that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED.

renewalSummary_updatedAt :: Lens' RenewalSummary UTCTime

The time at which the renewal summary was last updated.

ResourceRecord

resourceRecord_name :: Lens' ResourceRecord Text

The name of the DNS record to create in your domain. This is supplied by ACM.

resourceRecord_type :: Lens' ResourceRecord RecordType

The type of DNS record. Currently this can be CNAME.

resourceRecord_value :: Lens' ResourceRecord Text

The value of the CNAME record to add to your DNS database. This is supplied by ACM.

Tag

tag_value :: Lens' Tag (Maybe Text)

The value of the tag.

tag_key :: Lens' Tag Text

The key of the tag.