Network.AWS.CertificateManager.ResendValidationEmail

Description

Resends the email that requests domain ownership validation. The domain owner or an authorized representative must approve the ACM certificate before it can be issued. The certificate can be approved by clicking a link in the mail to navigate to the Amazon certificate approval website and then clicking I Approve. However, the validation email can be blocked by spam filters. Therefore, if you do not receive the original mail, you can request that the mail be resent within 72 hours of requesting the ACM certificate. If more than 72 hours have elapsed since your original request or since your last attempt to resend validation mail, you must request a new certificate. For more information about setting up your contact email addresses, see Configure Email for your Domain.

Creating a Request

data ResendValidationEmail

See: newResendValidationEmail smart constructor.

Constructors

ResendValidationEmail'

Fields

certificateArn :: Text

String that contains the ARN of the requested certificate. The certificate ARN is generated and returned by the RequestCertificate action as soon as the request is made. By default, using this parameter causes email to be sent to all top-level domains you specified in the certificate request. The ARN must be of the form:

arn:aws:acm:us-east-1:123456789012:certificate/12345678-1234-1234-1234-123456789012

domain :: Text

The fully qualified domain name (FQDN) of the certificate that needs to be validated.

validationDomain :: Text

The base validation domain that will act as the suffix of the email addresses that are used to send the emails. This must be the same as the Domain value or a superdomain of the Domain value. For example, if you requested a certificate for site.subdomain.example.com and specify a ValidationDomain of subdomain.example.com, ACM sends email to the domain registrant, technical contact, and administrative contact in WHOIS and the following five addresses:

admin@subdomain.example.com

administrator@subdomain.example.com

hostmaster@subdomain.example.com

postmaster@subdomain.example.com

webmaster@subdomain.example.com

newResendValidationEmail

Arguments

:: Text
$sel:certificateArn:ResendValidationEmail'

-> Text
$sel:domain:ResendValidationEmail'

-> Text
$sel:validationDomain:ResendValidationEmail'

-> ResendValidationEmail

Create a value of ResendValidationEmail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateArn:ResendValidationEmail', resendValidationEmail_certificateArn - String that contains the ARN of the requested certificate. The certificate ARN is generated and returned by the RequestCertificate action as soon as the request is made. By default, using this parameter causes email to be sent to all top-level domains you specified in the certificate request. The ARN must be of the form:

arn:aws:acm:us-east-1:123456789012:certificate/12345678-1234-1234-1234-123456789012

$sel:domain:ResendValidationEmail', resendValidationEmail_domain - The fully qualified domain name (FQDN) of the certificate that needs to be validated.

$sel:validationDomain:ResendValidationEmail', resendValidationEmail_validationDomain - The base validation domain that will act as the suffix of the email addresses that are used to send the emails. This must be the same as the Domain value or a superdomain of the Domain value. For example, if you requested a certificate for site.subdomain.example.com and specify a ValidationDomain of subdomain.example.com, ACM sends email to the domain registrant, technical contact, and administrative contact in WHOIS and the following five addresses:

admin@subdomain.example.com

administrator@subdomain.example.com

hostmaster@subdomain.example.com

postmaster@subdomain.example.com

webmaster@subdomain.example.com

Request Lenses

resendValidationEmail_certificateArn :: Lens' ResendValidationEmail Text

String that contains the ARN of the requested certificate. The certificate ARN is generated and returned by the RequestCertificate action as soon as the request is made. By default, using this parameter causes email to be sent to all top-level domains you specified in the certificate request. The ARN must be of the form:

arn:aws:acm:us-east-1:123456789012:certificate/12345678-1234-1234-1234-123456789012

resendValidationEmail_domain :: Lens' ResendValidationEmail Text

The fully qualified domain name (FQDN) of the certificate that needs to be validated.

resendValidationEmail_validationDomain :: Lens' ResendValidationEmail Text

The base validation domain that will act as the suffix of the email addresses that are used to send the emails. This must be the same as the Domain value or a superdomain of the Domain value. For example, if you requested a certificate for site.subdomain.example.com and specify a ValidationDomain of subdomain.example.com, ACM sends email to the domain registrant, technical contact, and administrative contact in WHOIS and the following five addresses:

admin@subdomain.example.com

administrator@subdomain.example.com

hostmaster@subdomain.example.com

postmaster@subdomain.example.com

webmaster@subdomain.example.com

Destructuring the Response

data ResendValidationEmailResponse

See: newResendValidationEmailResponse smart constructor.

Constructors

ResendValidationEmailResponse'

newResendValidationEmailResponse :: ResendValidationEmailResponse

Create a value of ResendValidationEmailResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.