Network.AWS.CertificateManager.Types

Description

Service Configuration

defaultService :: Service

API version 2015-12-08 of the Amazon Certificate Manager SDK configuration.

Errors

_TooManyTagsException :: AsError a => Getting (First ServiceError) a ServiceError

The request contains too many tags. Try the request again with fewer tags.

_InvalidStateException :: AsError a => Getting (First ServiceError) a ServiceError

Processing has reached an invalid state.

_InvalidArgsException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of of request parameters specified is not valid.

_InvalidArnException :: AsError a => Getting (First ServiceError) a ServiceError

The requested Amazon Resource Name (ARN) does not refer to an existing resource.

_TagPolicyException :: AsError a => Getting (First ServiceError) a ServiceError

A specified tag did not comply with an existing tag policy and was rejected.

_InvalidDomainValidationOptionsException :: AsError a => Getting (First ServiceError) a ServiceError

One or more values in the DomainValidationOption structure is incorrect.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied because it exceeded a quota.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

An input parameter was invalid.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have access required to perform this action.

_InvalidTagException :: AsError a => Getting (First ServiceError) a ServiceError

One or both of the values that make up the key-value pair is not valid. For example, you cannot specify a tag value that begins with aws:.

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The supplied input failed to satisfy constraints of an Amazon Web Services service.

_RequestInProgressException :: AsError a => Getting (First ServiceError) a ServiceError

The certificate request is in process and the certificate in your account has not yet been issued.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

You are trying to update a resource or configuration that is already being created or updated. Wait for the previous operation to finish and try again.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

An ACM quota has been exceeded.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The certificate is in use by another Amazon Web Services service in the caller's account. Remove the association and try again.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified certificate cannot be found in the caller's account or the caller's account cannot be found.

CertificateStatus

newtype CertificateStatus

Constructors

CertificateStatus'

Fields

fromCertificateStatus :: Text

Bundled Patterns

pattern CertificateStatus_EXPIRED :: CertificateStatus

pattern CertificateStatus_FAILED :: CertificateStatus

pattern CertificateStatus_INACTIVE :: CertificateStatus

pattern CertificateStatus_ISSUED :: CertificateStatus

pattern CertificateStatus_PENDING_VALIDATION :: CertificateStatus

pattern CertificateStatus_REVOKED :: CertificateStatus

pattern CertificateStatus_VALIDATION_TIMED_OUT :: CertificateStatus

CertificateTransparencyLoggingPreference

newtype CertificateTransparencyLoggingPreference

Constructors

CertificateTransparencyLoggingPreference'

Fields

fromCertificateTransparencyLoggingPreference :: Text

Bundled Patterns

pattern CertificateTransparencyLoggingPreference_DISABLED :: CertificateTransparencyLoggingPreference

pattern CertificateTransparencyLoggingPreference_ENABLED :: CertificateTransparencyLoggingPreference

CertificateType

newtype CertificateType

Constructors

CertificateType'

Fields

fromCertificateType :: Text

Bundled Patterns

pattern CertificateType_AMAZON_ISSUED :: CertificateType

pattern CertificateType_IMPORTED :: CertificateType

pattern CertificateType_PRIVATE :: CertificateType

DomainStatus

newtype DomainStatus

Constructors

DomainStatus'

Fields

fromDomainStatus :: Text

Bundled Patterns

pattern DomainStatus_FAILED :: DomainStatus

pattern DomainStatus_PENDING_VALIDATION :: DomainStatus

pattern DomainStatus_SUCCESS :: DomainStatus

ExtendedKeyUsageName

newtype ExtendedKeyUsageName

Constructors

ExtendedKeyUsageName'

Fields

fromExtendedKeyUsageName :: Text

Bundled Patterns

pattern ExtendedKeyUsageName_ANY :: ExtendedKeyUsageName

pattern ExtendedKeyUsageName_CODE_SIGNING :: ExtendedKeyUsageName

pattern ExtendedKeyUsageName_CUSTOM :: ExtendedKeyUsageName

pattern ExtendedKeyUsageName_EMAIL_PROTECTION :: ExtendedKeyUsageName

pattern ExtendedKeyUsageName_IPSEC_END_SYSTEM :: ExtendedKeyUsageName

pattern ExtendedKeyUsageName_IPSEC_TUNNEL :: ExtendedKeyUsageName

pattern ExtendedKeyUsageName_IPSEC_USER :: ExtendedKeyUsageName

pattern ExtendedKeyUsageName_NONE :: ExtendedKeyUsageName

pattern ExtendedKeyUsageName_OCSP_SIGNING :: ExtendedKeyUsageName

pattern ExtendedKeyUsageName_TIME_STAMPING :: ExtendedKeyUsageName

pattern ExtendedKeyUsageName_TLS_WEB_CLIENT_AUTHENTICATION :: ExtendedKeyUsageName

pattern ExtendedKeyUsageName_TLS_WEB_SERVER_AUTHENTICATION :: ExtendedKeyUsageName

FailureReason

newtype FailureReason

Constructors

FailureReason'

Fields

fromFailureReason :: Text

Bundled Patterns

pattern FailureReason_ADDITIONAL_VERIFICATION_REQUIRED :: FailureReason

pattern FailureReason_CAA_ERROR :: FailureReason

pattern FailureReason_DOMAIN_NOT_ALLOWED :: FailureReason

pattern FailureReason_DOMAIN_VALIDATION_DENIED :: FailureReason

pattern FailureReason_INVALID_PUBLIC_DOMAIN :: FailureReason

pattern FailureReason_NO_AVAILABLE_CONTACTS :: FailureReason

pattern FailureReason_OTHER :: FailureReason

pattern FailureReason_PCA_ACCESS_DENIED :: FailureReason

pattern FailureReason_PCA_INVALID_ARGS :: FailureReason

pattern FailureReason_PCA_INVALID_ARN :: FailureReason

pattern FailureReason_PCA_INVALID_DURATION :: FailureReason

pattern FailureReason_PCA_INVALID_STATE :: FailureReason

pattern FailureReason_PCA_LIMIT_EXCEEDED :: FailureReason

pattern FailureReason_PCA_NAME_CONSTRAINTS_VALIDATION :: FailureReason

pattern FailureReason_PCA_REQUEST_FAILED :: FailureReason

pattern FailureReason_PCA_RESOURCE_NOT_FOUND :: FailureReason

pattern FailureReason_SLR_NOT_FOUND :: FailureReason

KeyAlgorithm

newtype KeyAlgorithm

Constructors

KeyAlgorithm'

Fields

fromKeyAlgorithm :: Text

Bundled Patterns

pattern KeyAlgorithm_EC_prime256v1 :: KeyAlgorithm

pattern KeyAlgorithm_EC_secp384r1 :: KeyAlgorithm

pattern KeyAlgorithm_EC_secp521r1 :: KeyAlgorithm

pattern KeyAlgorithm_RSA_1024 :: KeyAlgorithm

pattern KeyAlgorithm_RSA_2048 :: KeyAlgorithm

pattern KeyAlgorithm_RSA_3072 :: KeyAlgorithm

pattern KeyAlgorithm_RSA_4096 :: KeyAlgorithm

KeyUsageName

newtype KeyUsageName

Constructors

KeyUsageName'

Fields

fromKeyUsageName :: Text

Bundled Patterns

pattern KeyUsageName_ANY :: KeyUsageName

pattern KeyUsageName_CERTIFICATE_SIGNING :: KeyUsageName

pattern KeyUsageName_CRL_SIGNING :: KeyUsageName

pattern KeyUsageName_CUSTOM :: KeyUsageName

pattern KeyUsageName_DATA_ENCIPHERMENT :: KeyUsageName

pattern KeyUsageName_DECIPHER_ONLY :: KeyUsageName

pattern KeyUsageName_DIGITAL_SIGNATURE :: KeyUsageName

pattern KeyUsageName_ENCIPHER_ONLY :: KeyUsageName

pattern KeyUsageName_KEY_AGREEMENT :: KeyUsageName

pattern KeyUsageName_KEY_ENCIPHERMENT :: KeyUsageName

pattern KeyUsageName_NON_REPUDIATION :: KeyUsageName

RecordType

newtype RecordType

Constructors

RecordType'

Fields

fromRecordType :: Text

Bundled Patterns

pattern RecordType_CNAME :: RecordType

RenewalEligibility

newtype RenewalEligibility

Constructors

RenewalEligibility'

Fields

fromRenewalEligibility :: Text

Bundled Patterns

pattern RenewalEligibility_ELIGIBLE :: RenewalEligibility

pattern RenewalEligibility_INELIGIBLE :: RenewalEligibility

RenewalStatus

newtype RenewalStatus

Constructors

RenewalStatus'

Fields

fromRenewalStatus :: Text

Bundled Patterns

pattern RenewalStatus_FAILED :: RenewalStatus

pattern RenewalStatus_PENDING_AUTO_RENEWAL :: RenewalStatus

pattern RenewalStatus_PENDING_VALIDATION :: RenewalStatus

pattern RenewalStatus_SUCCESS :: RenewalStatus

RevocationReason

newtype RevocationReason

Constructors

RevocationReason'

Fields

fromRevocationReason :: Text

Bundled Patterns

pattern RevocationReason_AFFILIATION_CHANGED :: RevocationReason

pattern RevocationReason_A_A_COMPROMISE :: RevocationReason

pattern RevocationReason_CA_COMPROMISE :: RevocationReason

pattern RevocationReason_CERTIFICATE_HOLD :: RevocationReason

pattern RevocationReason_CESSATION_OF_OPERATION :: RevocationReason

pattern RevocationReason_KEY_COMPROMISE :: RevocationReason

pattern RevocationReason_PRIVILEGE_WITHDRAWN :: RevocationReason

pattern RevocationReason_REMOVE_FROM_CRL :: RevocationReason

pattern RevocationReason_SUPERCEDED :: RevocationReason

pattern RevocationReason_UNSPECIFIED :: RevocationReason

ValidationMethod

newtype ValidationMethod

Constructors

ValidationMethod'

Fields

fromValidationMethod :: Text

Bundled Patterns

pattern ValidationMethod_DNS :: ValidationMethod

pattern ValidationMethod_EMAIL :: ValidationMethod

CertificateDetail

data CertificateDetail

Contains metadata about an ACM certificate. This structure is returned in the response to a DescribeCertificate request.

See: newCertificateDetail smart constructor.

Constructors

CertificateDetail'

Fields

status :: Maybe CertificateStatus

The status of the certificate.

notBefore :: Maybe POSIX

The time before which the certificate is not valid.

certificateAuthorityArn :: Maybe Text

The Amazon Resource Name (ARN) of the ACM PCA private certificate authority (CA) that issued the certificate. This has the following format:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

importedAt :: Maybe POSIX

The date and time at which the certificate was imported. This value exists only when the certificate type is IMPORTED.

extendedKeyUsages :: Maybe [ExtendedKeyUsage]

Contains a list of Extended Key Usage X.509 v3 extension objects. Each object specifies a purpose for which the certificate public key can be used and consists of a name and an object identifier (OID).

domainValidationOptions :: Maybe (NonEmpty DomainValidation)

Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED.

renewalEligibility :: Maybe RenewalEligibility

Specifies whether the certificate is eligible for renewal. At this time, only exported private certificates can be renewed with the RenewCertificate command.

options :: Maybe CertificateOptions

Value that specifies whether to add the certificate to a transparency log. Certificate transparency makes it possible to detect SSL certificates that have been mistakenly or maliciously issued. A browser might respond to certificate that has not been logged by showing an error message. The logs are cryptographically secure.

serial :: Maybe Text

The serial number of the certificate.

certificateArn :: Maybe Text

The Amazon Resource Name (ARN) of the certificate. For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

createdAt :: Maybe POSIX

The time at which the certificate was requested.

inUseBy :: Maybe [Text]

A list of ARNs for the Amazon Web Services resources that are using the certificate. A certificate can be used by multiple Amazon Web Services resources.

domainName :: Maybe Text

The fully qualified domain name for the certificate, such as www.example.com or example.com.

subjectAlternativeNames :: Maybe (NonEmpty Text)

One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate. The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

revocationReason :: Maybe RevocationReason

The reason the certificate was revoked. This value exists only when the certificate status is REVOKED.

failureReason :: Maybe FailureReason

The reason the certificate request failed. This value exists only when the certificate status is FAILED. For more information, see Certificate Request Failed in the Amazon Web Services Certificate Manager User Guide.

subject :: Maybe Text

The name of the entity that is associated with the public key contained in the certificate.

notAfter :: Maybe POSIX

The time after which the certificate is not valid.

revokedAt :: Maybe POSIX

The time at which the certificate was revoked. This value exists only when the certificate status is REVOKED.

keyUsages :: Maybe [KeyUsage]

A list of Key Usage X.509 v3 extension objects. Each object is a string value that identifies the purpose of the public key contained in the certificate. Possible extension values include DIGITAL_SIGNATURE, KEY_ENCHIPHERMENT, NON_REPUDIATION, and more.

issuer :: Maybe Text

The name of the certificate authority that issued and signed the certificate.

signatureAlgorithm :: Maybe Text

The algorithm that was used to sign the certificate.

issuedAt :: Maybe POSIX

The time at which the certificate was issued. This value exists only when the certificate type is AMAZON_ISSUED.

keyAlgorithm :: Maybe KeyAlgorithm

The algorithm that was used to generate the public-private key pair.

type' :: Maybe CertificateType

The source of the certificate. For certificates provided by ACM, this value is AMAZON_ISSUED. For certificates that you imported with ImportCertificate, this value is IMPORTED. ACM does not provide managed renewal for imported certificates. For more information about the differences between certificates that you import and those that ACM provides, see Importing Certificates in the Amazon Web Services Certificate Manager User Guide.

renewalSummary :: Maybe RenewalSummary

Contains information about the status of ACM's managed renewal for the certificate. This field exists only when the certificate type is AMAZON_ISSUED.

newCertificateDetail :: CertificateDetail

Create a value of CertificateDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CertificateDetail', certificateDetail_status - The status of the certificate.

$sel:notBefore:CertificateDetail', certificateDetail_notBefore - The time before which the certificate is not valid.

$sel:certificateAuthorityArn:CertificateDetail', certificateDetail_certificateAuthorityArn - The Amazon Resource Name (ARN) of the ACM PCA private certificate authority (CA) that issued the certificate. This has the following format:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

$sel:importedAt:CertificateDetail', certificateDetail_importedAt - The date and time at which the certificate was imported. This value exists only when the certificate type is IMPORTED.

$sel:extendedKeyUsages:CertificateDetail', certificateDetail_extendedKeyUsages - Contains a list of Extended Key Usage X.509 v3 extension objects. Each object specifies a purpose for which the certificate public key can be used and consists of a name and an object identifier (OID).

$sel:domainValidationOptions:CertificateDetail', certificateDetail_domainValidationOptions - Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED.

$sel:renewalEligibility:CertificateDetail', certificateDetail_renewalEligibility - Specifies whether the certificate is eligible for renewal. At this time, only exported private certificates can be renewed with the RenewCertificate command.

$sel:options:CertificateDetail', certificateDetail_options - Value that specifies whether to add the certificate to a transparency log. Certificate transparency makes it possible to detect SSL certificates that have been mistakenly or maliciously issued. A browser might respond to certificate that has not been logged by showing an error message. The logs are cryptographically secure.

$sel:serial:CertificateDetail', certificateDetail_serial - The serial number of the certificate.

$sel:certificateArn:CertificateDetail', certificateDetail_certificateArn - The Amazon Resource Name (ARN) of the certificate. For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

$sel:createdAt:CertificateDetail', certificateDetail_createdAt - The time at which the certificate was requested.

$sel:inUseBy:CertificateDetail', certificateDetail_inUseBy - A list of ARNs for the Amazon Web Services resources that are using the certificate. A certificate can be used by multiple Amazon Web Services resources.

$sel:domainName:CertificateDetail', certificateDetail_domainName - The fully qualified domain name for the certificate, such as www.example.com or example.com.

$sel:subjectAlternativeNames:CertificateDetail', certificateDetail_subjectAlternativeNames - One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate. The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

$sel:revocationReason:CertificateDetail', certificateDetail_revocationReason - The reason the certificate was revoked. This value exists only when the certificate status is REVOKED.

$sel:failureReason:CertificateDetail', certificateDetail_failureReason - The reason the certificate request failed. This value exists only when the certificate status is FAILED. For more information, see Certificate Request Failed in the Amazon Web Services Certificate Manager User Guide.

$sel:subject:CertificateDetail', certificateDetail_subject - The name of the entity that is associated with the public key contained in the certificate.

$sel:notAfter:CertificateDetail', certificateDetail_notAfter - The time after which the certificate is not valid.

$sel:revokedAt:CertificateDetail', certificateDetail_revokedAt - The time at which the certificate was revoked. This value exists only when the certificate status is REVOKED.

$sel:keyUsages:CertificateDetail', certificateDetail_keyUsages - A list of Key Usage X.509 v3 extension objects. Each object is a string value that identifies the purpose of the public key contained in the certificate. Possible extension values include DIGITAL_SIGNATURE, KEY_ENCHIPHERMENT, NON_REPUDIATION, and more.

$sel:issuer:CertificateDetail', certificateDetail_issuer - The name of the certificate authority that issued and signed the certificate.

$sel:signatureAlgorithm:CertificateDetail', certificateDetail_signatureAlgorithm - The algorithm that was used to sign the certificate.

$sel:issuedAt:CertificateDetail', certificateDetail_issuedAt - The time at which the certificate was issued. This value exists only when the certificate type is AMAZON_ISSUED.

$sel:keyAlgorithm:CertificateDetail', certificateDetail_keyAlgorithm - The algorithm that was used to generate the public-private key pair.

$sel:type':CertificateDetail', certificateDetail_type - The source of the certificate. For certificates provided by ACM, this value is AMAZON_ISSUED. For certificates that you imported with ImportCertificate, this value is IMPORTED. ACM does not provide managed renewal for imported certificates. For more information about the differences between certificates that you import and those that ACM provides, see Importing Certificates in the Amazon Web Services Certificate Manager User Guide.

$sel:renewalSummary:CertificateDetail', certificateDetail_renewalSummary - Contains information about the status of ACM's managed renewal for the certificate. This field exists only when the certificate type is AMAZON_ISSUED.

certificateDetail_status :: Lens' CertificateDetail (Maybe CertificateStatus)

The status of the certificate.

certificateDetail_notBefore :: Lens' CertificateDetail (Maybe UTCTime)

The time before which the certificate is not valid.

certificateDetail_certificateAuthorityArn :: Lens' CertificateDetail (Maybe Text)

The Amazon Resource Name (ARN) of the ACM PCA private certificate authority (CA) that issued the certificate. This has the following format:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

certificateDetail_importedAt :: Lens' CertificateDetail (Maybe UTCTime)

The date and time at which the certificate was imported. This value exists only when the certificate type is IMPORTED.

certificateDetail_extendedKeyUsages :: Lens' CertificateDetail (Maybe [ExtendedKeyUsage])

Contains a list of Extended Key Usage X.509 v3 extension objects. Each object specifies a purpose for which the certificate public key can be used and consists of a name and an object identifier (OID).

certificateDetail_domainValidationOptions :: Lens' CertificateDetail (Maybe (NonEmpty DomainValidation))

Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED.

certificateDetail_renewalEligibility :: Lens' CertificateDetail (Maybe RenewalEligibility)

Specifies whether the certificate is eligible for renewal. At this time, only exported private certificates can be renewed with the RenewCertificate command.

certificateDetail_options :: Lens' CertificateDetail (Maybe CertificateOptions)

Value that specifies whether to add the certificate to a transparency log. Certificate transparency makes it possible to detect SSL certificates that have been mistakenly or maliciously issued. A browser might respond to certificate that has not been logged by showing an error message. The logs are cryptographically secure.

certificateDetail_serial :: Lens' CertificateDetail (Maybe Text)

The serial number of the certificate.

certificateDetail_certificateArn :: Lens' CertificateDetail (Maybe Text)

The Amazon Resource Name (ARN) of the certificate. For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

certificateDetail_createdAt :: Lens' CertificateDetail (Maybe UTCTime)

The time at which the certificate was requested.

certificateDetail_inUseBy :: Lens' CertificateDetail (Maybe [Text])

A list of ARNs for the Amazon Web Services resources that are using the certificate. A certificate can be used by multiple Amazon Web Services resources.

certificateDetail_domainName :: Lens' CertificateDetail (Maybe Text)

The fully qualified domain name for the certificate, such as www.example.com or example.com.

certificateDetail_subjectAlternativeNames :: Lens' CertificateDetail (Maybe (NonEmpty Text))

One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate. The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

certificateDetail_revocationReason :: Lens' CertificateDetail (Maybe RevocationReason)

The reason the certificate was revoked. This value exists only when the certificate status is REVOKED.

certificateDetail_failureReason :: Lens' CertificateDetail (Maybe FailureReason)

The reason the certificate request failed. This value exists only when the certificate status is FAILED. For more information, see Certificate Request Failed in the Amazon Web Services Certificate Manager User Guide.

certificateDetail_subject :: Lens' CertificateDetail (Maybe Text)

The name of the entity that is associated with the public key contained in the certificate.

certificateDetail_notAfter :: Lens' CertificateDetail (Maybe UTCTime)

The time after which the certificate is not valid.

certificateDetail_revokedAt :: Lens' CertificateDetail (Maybe UTCTime)

The time at which the certificate was revoked. This value exists only when the certificate status is REVOKED.

certificateDetail_keyUsages :: Lens' CertificateDetail (Maybe [KeyUsage])

A list of Key Usage X.509 v3 extension objects. Each object is a string value that identifies the purpose of the public key contained in the certificate. Possible extension values include DIGITAL_SIGNATURE, KEY_ENCHIPHERMENT, NON_REPUDIATION, and more.

certificateDetail_issuer :: Lens' CertificateDetail (Maybe Text)

The name of the certificate authority that issued and signed the certificate.

certificateDetail_signatureAlgorithm :: Lens' CertificateDetail (Maybe Text)

The algorithm that was used to sign the certificate.

certificateDetail_issuedAt :: Lens' CertificateDetail (Maybe UTCTime)

The time at which the certificate was issued. This value exists only when the certificate type is AMAZON_ISSUED.

certificateDetail_keyAlgorithm :: Lens' CertificateDetail (Maybe KeyAlgorithm)

The algorithm that was used to generate the public-private key pair.

certificateDetail_type :: Lens' CertificateDetail (Maybe CertificateType)

The source of the certificate. For certificates provided by ACM, this value is AMAZON_ISSUED. For certificates that you imported with ImportCertificate, this value is IMPORTED. ACM does not provide managed renewal for imported certificates. For more information about the differences between certificates that you import and those that ACM provides, see Importing Certificates in the Amazon Web Services Certificate Manager User Guide.

certificateDetail_renewalSummary :: Lens' CertificateDetail (Maybe RenewalSummary)

Contains information about the status of ACM's managed renewal for the certificate. This field exists only when the certificate type is AMAZON_ISSUED.

CertificateOptions

data CertificateOptions

Structure that contains options for your certificate. Currently, you can use this only to specify whether to opt in to or out of certificate transparency logging. Some browsers require that public certificates issued for your domain be recorded in a log. Certificates that are not logged typically generate a browser error. Transparency makes it possible for you to detect SSL/TLS certificates that have been mistakenly or maliciously issued for your domain. For general information, see Certificate Transparency Logging.

See: newCertificateOptions smart constructor.

Constructors

CertificateOptions'

Fields

certificateTransparencyLoggingPreference :: Maybe CertificateTransparencyLoggingPreference

You can opt out of certificate transparency logging by specifying the DISABLED option. Opt in by specifying ENABLED.

newCertificateOptions :: CertificateOptions

Create a value of CertificateOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateTransparencyLoggingPreference:CertificateOptions', certificateOptions_certificateTransparencyLoggingPreference - You can opt out of certificate transparency logging by specifying the DISABLED option. Opt in by specifying ENABLED.

certificateOptions_certificateTransparencyLoggingPreference :: Lens' CertificateOptions (Maybe CertificateTransparencyLoggingPreference)

You can opt out of certificate transparency logging by specifying the DISABLED option. Opt in by specifying ENABLED.

CertificateSummary

data CertificateSummary

This structure is returned in the response object of ListCertificates action.

See: newCertificateSummary smart constructor.

Constructors

CertificateSummary'

Fields

certificateArn :: Maybe Text

Amazon Resource Name (ARN) of the certificate. This is of the form:

arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012

For more information about ARNs, see Amazon Resource Names (ARNs).

domainName :: Maybe Text

Fully qualified domain name (FQDN), such as www.example.com or example.com, for the certificate.

newCertificateSummary :: CertificateSummary

Create a value of CertificateSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateArn:CertificateSummary', certificateSummary_certificateArn - Amazon Resource Name (ARN) of the certificate. This is of the form:

arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012

For more information about ARNs, see Amazon Resource Names (ARNs).

$sel:domainName:CertificateSummary', certificateSummary_domainName - Fully qualified domain name (FQDN), such as www.example.com or example.com, for the certificate.

certificateSummary_certificateArn :: Lens' CertificateSummary (Maybe Text)

Amazon Resource Name (ARN) of the certificate. This is of the form:

arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012

For more information about ARNs, see Amazon Resource Names (ARNs).

certificateSummary_domainName :: Lens' CertificateSummary (Maybe Text)

Fully qualified domain name (FQDN), such as www.example.com or example.com, for the certificate.

DomainValidation

data DomainValidation

Contains information about the validation of each domain name in the certificate.

See: newDomainValidation smart constructor.

Constructors

DomainValidation'

Fields

resourceRecord :: Maybe ResourceRecord

Contains the CNAME record that you add to your DNS database for domain validation. For more information, see Use DNS to Validate Domain Ownership.

Note: The CNAME information that you need does not include the name of your domain. If you include your domain name in the DNS database CNAME record, validation fails. For example, if the name is "_a79865eb4cd1a6ab990a45779b4e0b96.yourdomain.com", only "_a79865eb4cd1a6ab990a45779b4e0b96" must be used.

validationMethod :: Maybe ValidationMethod

Specifies the domain validation method.

validationEmails :: Maybe [Text]

A list of email addresses that ACM used to send domain validation emails.

validationStatus :: Maybe DomainStatus

The validation status of the domain name. This can be one of the following values:

PENDING_VALIDATION

@@SUCCESS

@@FAILED

validationDomain :: Maybe Text

The domain name that ACM used to send domain validation emails.

domainName :: Text

A fully qualified domain name (FQDN) in the certificate. For example, www.example.com or example.com.

newDomainValidation

Arguments

:: Text
$sel:domainName:DomainValidation'

-> DomainValidation

Create a value of DomainValidation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceRecord:DomainValidation', domainValidation_resourceRecord - Contains the CNAME record that you add to your DNS database for domain validation. For more information, see Use DNS to Validate Domain Ownership.

Note: The CNAME information that you need does not include the name of your domain. If you include your domain name in the DNS database CNAME record, validation fails. For example, if the name is "_a79865eb4cd1a6ab990a45779b4e0b96.yourdomain.com", only "_a79865eb4cd1a6ab990a45779b4e0b96" must be used.

$sel:validationMethod:DomainValidation', domainValidation_validationMethod - Specifies the domain validation method.

$sel:validationEmails:DomainValidation', domainValidation_validationEmails - A list of email addresses that ACM used to send domain validation emails.

$sel:validationStatus:DomainValidation', domainValidation_validationStatus - The validation status of the domain name. This can be one of the following values:

PENDING_VALIDATION

@@SUCCESS

@@FAILED

$sel:validationDomain:DomainValidation', domainValidation_validationDomain - The domain name that ACM used to send domain validation emails.

$sel:domainName:DomainValidation', domainValidation_domainName - A fully qualified domain name (FQDN) in the certificate. For example, www.example.com or example.com.

domainValidation_resourceRecord :: Lens' DomainValidation (Maybe ResourceRecord)

Contains the CNAME record that you add to your DNS database for domain validation. For more information, see Use DNS to Validate Domain Ownership.

Note: The CNAME information that you need does not include the name of your domain. If you include your domain name in the DNS database CNAME record, validation fails. For example, if the name is "_a79865eb4cd1a6ab990a45779b4e0b96.yourdomain.com", only "_a79865eb4cd1a6ab990a45779b4e0b96" must be used.

domainValidation_validationMethod :: Lens' DomainValidation (Maybe ValidationMethod)

Specifies the domain validation method.

domainValidation_validationEmails :: Lens' DomainValidation (Maybe [Text])

A list of email addresses that ACM used to send domain validation emails.

domainValidation_validationStatus :: Lens' DomainValidation (Maybe DomainStatus)

The validation status of the domain name. This can be one of the following values:

PENDING_VALIDATION

@@SUCCESS

@@FAILED

domainValidation_validationDomain :: Lens' DomainValidation (Maybe Text)

The domain name that ACM used to send domain validation emails.

domainValidation_domainName :: Lens' DomainValidation Text

A fully qualified domain name (FQDN) in the certificate. For example, www.example.com or example.com.

DomainValidationOption

data DomainValidationOption

Contains information about the domain names that you want ACM to use to send you emails that enable you to validate domain ownership.

See: newDomainValidationOption smart constructor.

Constructors

DomainValidationOption'

Fields

domainName :: Text

A fully qualified domain name (FQDN) in the certificate request.

validationDomain :: Text

The domain name that you want ACM to use to send you validation emails. This domain name is the suffix of the email addresses that you want ACM to use. This must be the same as the DomainName value or a superdomain of the DomainName value. For example, if you request a certificate for testing.example.com, you can specify example.com for this value. In that case, ACM sends domain validation emails to the following five addresses:

admin@example.com

administrator@example.com

hostmaster@example.com

postmaster@example.com

webmaster@example.com

newDomainValidationOption

Arguments

:: Text
$sel:domainName:DomainValidationOption'

-> Text
$sel:validationDomain:DomainValidationOption'

-> DomainValidationOption

Create a value of DomainValidationOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:DomainValidationOption', domainValidationOption_domainName - A fully qualified domain name (FQDN) in the certificate request.

$sel:validationDomain:DomainValidationOption', domainValidationOption_validationDomain - The domain name that you want ACM to use to send you validation emails. This domain name is the suffix of the email addresses that you want ACM to use. This must be the same as the DomainName value or a superdomain of the DomainName value. For example, if you request a certificate for testing.example.com, you can specify example.com for this value. In that case, ACM sends domain validation emails to the following five addresses:

admin@example.com

administrator@example.com

hostmaster@example.com

postmaster@example.com

webmaster@example.com

domainValidationOption_domainName :: Lens' DomainValidationOption Text

A fully qualified domain name (FQDN) in the certificate request.

domainValidationOption_validationDomain :: Lens' DomainValidationOption Text

The domain name that you want ACM to use to send you validation emails. This domain name is the suffix of the email addresses that you want ACM to use. This must be the same as the DomainName value or a superdomain of the DomainName value. For example, if you request a certificate for testing.example.com, you can specify example.com for this value. In that case, ACM sends domain validation emails to the following five addresses:

admin@example.com

administrator@example.com

hostmaster@example.com

postmaster@example.com

webmaster@example.com

ExpiryEventsConfiguration

data ExpiryEventsConfiguration

Object containing expiration events options associated with an Amazon Web Services account.

See: newExpiryEventsConfiguration smart constructor.

Constructors

ExpiryEventsConfiguration'

Fields

daysBeforeExpiry :: Maybe Natural

Specifies the number of days prior to certificate expiration when ACM starts generating EventBridge events. ACM sends one event per day per certificate until the certificate expires. By default, accounts receive events starting 45 days before certificate expiration.

newExpiryEventsConfiguration :: ExpiryEventsConfiguration

Create a value of ExpiryEventsConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:daysBeforeExpiry:ExpiryEventsConfiguration', expiryEventsConfiguration_daysBeforeExpiry - Specifies the number of days prior to certificate expiration when ACM starts generating EventBridge events. ACM sends one event per day per certificate until the certificate expires. By default, accounts receive events starting 45 days before certificate expiration.

expiryEventsConfiguration_daysBeforeExpiry :: Lens' ExpiryEventsConfiguration (Maybe Natural)

Specifies the number of days prior to certificate expiration when ACM starts generating EventBridge events. ACM sends one event per day per certificate until the certificate expires. By default, accounts receive events starting 45 days before certificate expiration.

ExtendedKeyUsage

data ExtendedKeyUsage

The Extended Key Usage X.509 v3 extension defines one or more purposes for which the public key can be used. This is in addition to or in place of the basic purposes specified by the Key Usage extension.

See: newExtendedKeyUsage smart constructor.

Constructors

ExtendedKeyUsage'

Fields

name :: Maybe ExtendedKeyUsageName

The name of an Extended Key Usage value.

oid :: Maybe Text

An object identifier (OID) for the extension value. OIDs are strings of numbers separated by periods. The following OIDs are defined in RFC 3280 and RFC 5280.

1.3.6.1.5.5.7.3.1 (TLS_WEB_SERVER_AUTHENTICATION)

1.3.6.1.5.5.7.3.2 (TLS_WEB_CLIENT_AUTHENTICATION)

1.3.6.1.5.5.7.3.3 (CODE_SIGNING)

1.3.6.1.5.5.7.3.4 (EMAIL_PROTECTION)

1.3.6.1.5.5.7.3.8 (TIME_STAMPING)

1.3.6.1.5.5.7.3.9 (OCSP_SIGNING)

1.3.6.1.5.5.7.3.5 (IPSEC_END_SYSTEM)

1.3.6.1.5.5.7.3.6 (IPSEC_TUNNEL)

1.3.6.1.5.5.7.3.7 (IPSEC_USER)

newExtendedKeyUsage :: ExtendedKeyUsage

Create a value of ExtendedKeyUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ExtendedKeyUsage', extendedKeyUsage_name - The name of an Extended Key Usage value.

$sel:oid:ExtendedKeyUsage', extendedKeyUsage_oid - An object identifier (OID) for the extension value. OIDs are strings of numbers separated by periods. The following OIDs are defined in RFC 3280 and RFC 5280.

1.3.6.1.5.5.7.3.1 (TLS_WEB_SERVER_AUTHENTICATION)

1.3.6.1.5.5.7.3.2 (TLS_WEB_CLIENT_AUTHENTICATION)

1.3.6.1.5.5.7.3.3 (CODE_SIGNING)

1.3.6.1.5.5.7.3.4 (EMAIL_PROTECTION)

1.3.6.1.5.5.7.3.8 (TIME_STAMPING)

1.3.6.1.5.5.7.3.9 (OCSP_SIGNING)

1.3.6.1.5.5.7.3.5 (IPSEC_END_SYSTEM)

1.3.6.1.5.5.7.3.6 (IPSEC_TUNNEL)

1.3.6.1.5.5.7.3.7 (IPSEC_USER)

extendedKeyUsage_name :: Lens' ExtendedKeyUsage (Maybe ExtendedKeyUsageName)

The name of an Extended Key Usage value.

extendedKeyUsage_oid :: Lens' ExtendedKeyUsage (Maybe Text)

An object identifier (OID) for the extension value. OIDs are strings of numbers separated by periods. The following OIDs are defined in RFC 3280 and RFC 5280.

1.3.6.1.5.5.7.3.1 (TLS_WEB_SERVER_AUTHENTICATION)

1.3.6.1.5.5.7.3.2 (TLS_WEB_CLIENT_AUTHENTICATION)

1.3.6.1.5.5.7.3.3 (CODE_SIGNING)

1.3.6.1.5.5.7.3.4 (EMAIL_PROTECTION)

1.3.6.1.5.5.7.3.8 (TIME_STAMPING)

1.3.6.1.5.5.7.3.9 (OCSP_SIGNING)

1.3.6.1.5.5.7.3.5 (IPSEC_END_SYSTEM)

1.3.6.1.5.5.7.3.6 (IPSEC_TUNNEL)

1.3.6.1.5.5.7.3.7 (IPSEC_USER)

Filters

data Filters

This structure can be used in the ListCertificates action to filter the output of the certificate list.

See: newFilters smart constructor.

Constructors

Filters'

Fields

keyTypes :: Maybe [KeyAlgorithm]

Specify one or more algorithms that can be used to generate key pairs.

Default filtering returns only RSA_1024 and RSA_2048 certificates that have at least one domain. To return other certificate types, provide the desired type signatures in a comma-separated list. For example, "keyTypes": ["RSA_2048,RSA_4096"] returns both RSA_2048 and RSA_4096 certificates.

extendedKeyUsage :: Maybe [ExtendedKeyUsageName]

Specify one or more ExtendedKeyUsage extension values.

keyUsage :: Maybe [KeyUsageName]

Specify one or more KeyUsage extension values.

newFilters :: Filters

Create a value of Filters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyTypes:Filters', filters_keyTypes - Specify one or more algorithms that can be used to generate key pairs.

Default filtering returns only RSA_1024 and RSA_2048 certificates that have at least one domain. To return other certificate types, provide the desired type signatures in a comma-separated list. For example, "keyTypes": ["RSA_2048,RSA_4096"] returns both RSA_2048 and RSA_4096 certificates.

$sel:extendedKeyUsage:Filters', filters_extendedKeyUsage - Specify one or more ExtendedKeyUsage extension values.

$sel:keyUsage:Filters', filters_keyUsage - Specify one or more KeyUsage extension values.

filters_keyTypes :: Lens' Filters (Maybe [KeyAlgorithm])

Specify one or more algorithms that can be used to generate key pairs.

Default filtering returns only RSA_1024 and RSA_2048 certificates that have at least one domain. To return other certificate types, provide the desired type signatures in a comma-separated list. For example, "keyTypes": ["RSA_2048,RSA_4096"] returns both RSA_2048 and RSA_4096 certificates.

filters_extendedKeyUsage :: Lens' Filters (Maybe [ExtendedKeyUsageName])

Specify one or more ExtendedKeyUsage extension values.

filters_keyUsage :: Lens' Filters (Maybe [KeyUsageName])

Specify one or more KeyUsage extension values.

KeyUsage

data KeyUsage

The Key Usage X.509 v3 extension defines the purpose of the public key contained in the certificate.

See: newKeyUsage smart constructor.

Constructors

KeyUsage'

Fields

name :: Maybe KeyUsageName

A string value that contains a Key Usage extension name.

newKeyUsage :: KeyUsage

Create a value of KeyUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:KeyUsage', keyUsage_name - A string value that contains a Key Usage extension name.

keyUsage_name :: Lens' KeyUsage (Maybe KeyUsageName)

A string value that contains a Key Usage extension name.

RenewalSummary

data RenewalSummary

Contains information about the status of ACM's managed renewal for the certificate. This structure exists only when the certificate type is AMAZON_ISSUED.

See: newRenewalSummary smart constructor.

Constructors

RenewalSummary'

Fields

renewalStatusReason :: Maybe FailureReason

The reason that a renewal request was unsuccessful.

renewalStatus :: RenewalStatus

The status of ACM's managed renewal of the certificate.

domainValidationOptions :: NonEmpty DomainValidation

Contains information about the validation of each domain name in the certificate, as it pertains to ACM's managed renewal. This is different from the initial validation that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED.

updatedAt :: POSIX

The time at which the renewal summary was last updated.

newRenewalSummary

Arguments

:: RenewalStatus
$sel:renewalStatus:RenewalSummary'

-> NonEmpty DomainValidation
$sel:domainValidationOptions:RenewalSummary'

-> UTCTime
$sel:updatedAt:RenewalSummary'

-> RenewalSummary

Create a value of RenewalSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renewalStatusReason:RenewalSummary', renewalSummary_renewalStatusReason - The reason that a renewal request was unsuccessful.

$sel:renewalStatus:RenewalSummary', renewalSummary_renewalStatus - The status of ACM's managed renewal of the certificate.

$sel:domainValidationOptions:RenewalSummary', renewalSummary_domainValidationOptions - Contains information about the validation of each domain name in the certificate, as it pertains to ACM's managed renewal. This is different from the initial validation that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED.

$sel:updatedAt:RenewalSummary', renewalSummary_updatedAt - The time at which the renewal summary was last updated.

renewalSummary_renewalStatusReason :: Lens' RenewalSummary (Maybe FailureReason)

The reason that a renewal request was unsuccessful.

renewalSummary_renewalStatus :: Lens' RenewalSummary RenewalStatus

The status of ACM's managed renewal of the certificate.

renewalSummary_domainValidationOptions :: Lens' RenewalSummary (NonEmpty DomainValidation)

Contains information about the validation of each domain name in the certificate, as it pertains to ACM's managed renewal. This is different from the initial validation that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED.

renewalSummary_updatedAt :: Lens' RenewalSummary UTCTime

The time at which the renewal summary was last updated.

ResourceRecord

data ResourceRecord

Contains a DNS record value that you can use to validate ownership or control of a domain. This is used by the DescribeCertificate action.

See: newResourceRecord smart constructor.

Constructors

ResourceRecord'

Fields

name :: Text

The name of the DNS record to create in your domain. This is supplied by ACM.

type' :: RecordType

The type of DNS record. Currently this can be CNAME.

value :: Text

The value of the CNAME record to add to your DNS database. This is supplied by ACM.

newResourceRecord

Arguments

:: Text
$sel:name:ResourceRecord'

-> RecordType
$sel:type':ResourceRecord'

-> Text
$sel:value:ResourceRecord'

-> ResourceRecord

Create a value of ResourceRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ResourceRecord', resourceRecord_name - The name of the DNS record to create in your domain. This is supplied by ACM.

$sel:type':ResourceRecord', resourceRecord_type - The type of DNS record. Currently this can be CNAME.

$sel:value:ResourceRecord', resourceRecord_value - The value of the CNAME record to add to your DNS database. This is supplied by ACM.

resourceRecord_name :: Lens' ResourceRecord Text

The name of the DNS record to create in your domain. This is supplied by ACM.

resourceRecord_type :: Lens' ResourceRecord RecordType

The type of DNS record. Currently this can be CNAME.

resourceRecord_value :: Lens' ResourceRecord Text

The value of the CNAME record to add to your DNS database. This is supplied by ACM.

Tag

data Tag

A key-value pair that identifies or specifies metadata about an ACM resource.

See: newTag smart constructor.

Constructors

Tag'

Fields

value :: Maybe Text

The value of the tag.

key :: Text

The key of the tag.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - The value of the tag.

$sel:key:Tag', tag_key - The key of the tag.

tag_value :: Lens' Tag (Maybe Text)

The value of the tag.

tag_key :: Lens' Tag Text

The key of the tag.