Network.AWS.CertificateManager.Types.CertificateDetail

Description

Documentation

data CertificateDetail

Contains metadata about an ACM certificate. This structure is returned in the response to a DescribeCertificate request.

See: newCertificateDetail smart constructor.

Constructors

CertificateDetail'

Fields

subject :: Maybe Text

The name of the entity that is associated with the public key contained in the certificate.

status :: Maybe CertificateStatus

The status of the certificate.

failureReason :: Maybe FailureReason

The reason the certificate request failed. This value exists only when the certificate status is FAILED. For more information, see Certificate Request Failed in the Amazon Web Services Certificate Manager User Guide.

subjectAlternativeNames :: Maybe (NonEmpty Text)

One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate. The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

inUseBy :: Maybe [Text]

A list of ARNs for the Amazon Web Services resources that are using the certificate. A certificate can be used by multiple Amazon Web Services resources.

createdAt :: Maybe POSIX

The time at which the certificate was requested.

certificateArn :: Maybe Text

The Amazon Resource Name (ARN) of the certificate. For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

serial :: Maybe Text

The serial number of the certificate.

renewalEligibility :: Maybe RenewalEligibility

Specifies whether the certificate is eligible for renewal. At this time, only exported private certificates can be renewed with the RenewCertificate command.

extendedKeyUsages :: Maybe [ExtendedKeyUsage]

Contains a list of Extended Key Usage X.509 v3 extension objects. Each object specifies a purpose for which the certificate public key can be used and consists of a name and an object identifier (OID).

importedAt :: Maybe POSIX

The date and time at which the certificate was imported. This value exists only when the certificate type is IMPORTED.

keyUsages :: Maybe [KeyUsage]

A list of Key Usage X.509 v3 extension objects. Each object is a string value that identifies the purpose of the public key contained in the certificate. Possible extension values include DIGITAL_SIGNATURE, KEY_ENCHIPHERMENT, NON_REPUDIATION, and more.

revokedAt :: Maybe POSIX

The time at which the certificate was revoked. This value exists only when the certificate status is REVOKED.

notBefore :: Maybe POSIX

The time before which the certificate is not valid.

revocationReason :: Maybe RevocationReason

The reason the certificate was revoked. This value exists only when the certificate status is REVOKED.

domainName :: Maybe Text

The fully qualified domain name for the certificate, such as www.example.com or example.com.

renewalSummary :: Maybe RenewalSummary

Contains information about the status of ACM's managed renewal for the certificate. This field exists only when the certificate type is AMAZON_ISSUED.

keyAlgorithm :: Maybe KeyAlgorithm

The algorithm that was used to generate the public-private key pair.

type' :: Maybe CertificateType

The source of the certificate. For certificates provided by ACM, this value is AMAZON_ISSUED. For certificates that you imported with ImportCertificate, this value is IMPORTED. ACM does not provide managed renewal for imported certificates. For more information about the differences between certificates that you import and those that ACM provides, see Importing Certificates in the Amazon Web Services Certificate Manager User Guide.

options :: Maybe CertificateOptions

Value that specifies whether to add the certificate to a transparency log. Certificate transparency makes it possible to detect SSL certificates that have been mistakenly or maliciously issued. A browser might respond to certificate that has not been logged by showing an error message. The logs are cryptographically secure.

issuedAt :: Maybe POSIX

The time at which the certificate was issued. This value exists only when the certificate type is AMAZON_ISSUED.

signatureAlgorithm :: Maybe Text

The algorithm that was used to sign the certificate.

domainValidationOptions :: Maybe (NonEmpty DomainValidation)

Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED.

issuer :: Maybe Text

The name of the certificate authority that issued and signed the certificate.

notAfter :: Maybe POSIX

The time after which the certificate is not valid.

certificateAuthorityArn :: Maybe Text

The Amazon Resource Name (ARN) of the ACM PCA private certificate authority (CA) that issued the certificate. This has the following format:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

newCertificateDetail :: CertificateDetail

Create a value of CertificateDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subject:CertificateDetail', certificateDetail_subject - The name of the entity that is associated with the public key contained in the certificate.

$sel:status:CertificateDetail', certificateDetail_status - The status of the certificate.

$sel:failureReason:CertificateDetail', certificateDetail_failureReason - The reason the certificate request failed. This value exists only when the certificate status is FAILED. For more information, see Certificate Request Failed in the Amazon Web Services Certificate Manager User Guide.

$sel:subjectAlternativeNames:CertificateDetail', certificateDetail_subjectAlternativeNames - One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate. The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

$sel:inUseBy:CertificateDetail', certificateDetail_inUseBy - A list of ARNs for the Amazon Web Services resources that are using the certificate. A certificate can be used by multiple Amazon Web Services resources.

$sel:createdAt:CertificateDetail', certificateDetail_createdAt - The time at which the certificate was requested.

$sel:certificateArn:CertificateDetail', certificateDetail_certificateArn - The Amazon Resource Name (ARN) of the certificate. For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

$sel:serial:CertificateDetail', certificateDetail_serial - The serial number of the certificate.

$sel:renewalEligibility:CertificateDetail', certificateDetail_renewalEligibility - Specifies whether the certificate is eligible for renewal. At this time, only exported private certificates can be renewed with the RenewCertificate command.

$sel:extendedKeyUsages:CertificateDetail', certificateDetail_extendedKeyUsages - Contains a list of Extended Key Usage X.509 v3 extension objects. Each object specifies a purpose for which the certificate public key can be used and consists of a name and an object identifier (OID).

$sel:importedAt:CertificateDetail', certificateDetail_importedAt - The date and time at which the certificate was imported. This value exists only when the certificate type is IMPORTED.

$sel:keyUsages:CertificateDetail', certificateDetail_keyUsages - A list of Key Usage X.509 v3 extension objects. Each object is a string value that identifies the purpose of the public key contained in the certificate. Possible extension values include DIGITAL_SIGNATURE, KEY_ENCHIPHERMENT, NON_REPUDIATION, and more.

$sel:revokedAt:CertificateDetail', certificateDetail_revokedAt - The time at which the certificate was revoked. This value exists only when the certificate status is REVOKED.

$sel:notBefore:CertificateDetail', certificateDetail_notBefore - The time before which the certificate is not valid.

$sel:revocationReason:CertificateDetail', certificateDetail_revocationReason - The reason the certificate was revoked. This value exists only when the certificate status is REVOKED.

$sel:domainName:CertificateDetail', certificateDetail_domainName - The fully qualified domain name for the certificate, such as www.example.com or example.com.

$sel:renewalSummary:CertificateDetail', certificateDetail_renewalSummary - Contains information about the status of ACM's managed renewal for the certificate. This field exists only when the certificate type is AMAZON_ISSUED.

$sel:keyAlgorithm:CertificateDetail', certificateDetail_keyAlgorithm - The algorithm that was used to generate the public-private key pair.

$sel:type':CertificateDetail', certificateDetail_type - The source of the certificate. For certificates provided by ACM, this value is AMAZON_ISSUED. For certificates that you imported with ImportCertificate, this value is IMPORTED. ACM does not provide managed renewal for imported certificates. For more information about the differences between certificates that you import and those that ACM provides, see Importing Certificates in the Amazon Web Services Certificate Manager User Guide.

$sel:options:CertificateDetail', certificateDetail_options - Value that specifies whether to add the certificate to a transparency log. Certificate transparency makes it possible to detect SSL certificates that have been mistakenly or maliciously issued. A browser might respond to certificate that has not been logged by showing an error message. The logs are cryptographically secure.

$sel:issuedAt:CertificateDetail', certificateDetail_issuedAt - The time at which the certificate was issued. This value exists only when the certificate type is AMAZON_ISSUED.

$sel:signatureAlgorithm:CertificateDetail', certificateDetail_signatureAlgorithm - The algorithm that was used to sign the certificate.

$sel:domainValidationOptions:CertificateDetail', certificateDetail_domainValidationOptions - Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED.

$sel:issuer:CertificateDetail', certificateDetail_issuer - The name of the certificate authority that issued and signed the certificate.

$sel:notAfter:CertificateDetail', certificateDetail_notAfter - The time after which the certificate is not valid.

$sel:certificateAuthorityArn:CertificateDetail', certificateDetail_certificateAuthorityArn - The Amazon Resource Name (ARN) of the ACM PCA private certificate authority (CA) that issued the certificate. This has the following format:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

certificateDetail_subject :: Lens' CertificateDetail (Maybe Text)

The name of the entity that is associated with the public key contained in the certificate.

certificateDetail_status :: Lens' CertificateDetail (Maybe CertificateStatus)

The status of the certificate.

certificateDetail_failureReason :: Lens' CertificateDetail (Maybe FailureReason)

The reason the certificate request failed. This value exists only when the certificate status is FAILED. For more information, see Certificate Request Failed in the Amazon Web Services Certificate Manager User Guide.

certificateDetail_subjectAlternativeNames :: Lens' CertificateDetail (Maybe (NonEmpty Text))

One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate. The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

certificateDetail_inUseBy :: Lens' CertificateDetail (Maybe [Text])

A list of ARNs for the Amazon Web Services resources that are using the certificate. A certificate can be used by multiple Amazon Web Services resources.

certificateDetail_createdAt :: Lens' CertificateDetail (Maybe UTCTime)

The time at which the certificate was requested.

certificateDetail_certificateArn :: Lens' CertificateDetail (Maybe Text)

The Amazon Resource Name (ARN) of the certificate. For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

certificateDetail_serial :: Lens' CertificateDetail (Maybe Text)

The serial number of the certificate.

certificateDetail_renewalEligibility :: Lens' CertificateDetail (Maybe RenewalEligibility)

Specifies whether the certificate is eligible for renewal. At this time, only exported private certificates can be renewed with the RenewCertificate command.

certificateDetail_extendedKeyUsages :: Lens' CertificateDetail (Maybe [ExtendedKeyUsage])

Contains a list of Extended Key Usage X.509 v3 extension objects. Each object specifies a purpose for which the certificate public key can be used and consists of a name and an object identifier (OID).

certificateDetail_importedAt :: Lens' CertificateDetail (Maybe UTCTime)

The date and time at which the certificate was imported. This value exists only when the certificate type is IMPORTED.

certificateDetail_keyUsages :: Lens' CertificateDetail (Maybe [KeyUsage])

A list of Key Usage X.509 v3 extension objects. Each object is a string value that identifies the purpose of the public key contained in the certificate. Possible extension values include DIGITAL_SIGNATURE, KEY_ENCHIPHERMENT, NON_REPUDIATION, and more.

certificateDetail_revokedAt :: Lens' CertificateDetail (Maybe UTCTime)

The time at which the certificate was revoked. This value exists only when the certificate status is REVOKED.

certificateDetail_notBefore :: Lens' CertificateDetail (Maybe UTCTime)

The time before which the certificate is not valid.

certificateDetail_revocationReason :: Lens' CertificateDetail (Maybe RevocationReason)

The reason the certificate was revoked. This value exists only when the certificate status is REVOKED.

certificateDetail_domainName :: Lens' CertificateDetail (Maybe Text)

The fully qualified domain name for the certificate, such as www.example.com or example.com.

certificateDetail_renewalSummary :: Lens' CertificateDetail (Maybe RenewalSummary)

Contains information about the status of ACM's managed renewal for the certificate. This field exists only when the certificate type is AMAZON_ISSUED.

certificateDetail_keyAlgorithm :: Lens' CertificateDetail (Maybe KeyAlgorithm)

The algorithm that was used to generate the public-private key pair.

certificateDetail_type :: Lens' CertificateDetail (Maybe CertificateType)

The source of the certificate. For certificates provided by ACM, this value is AMAZON_ISSUED. For certificates that you imported with ImportCertificate, this value is IMPORTED. ACM does not provide managed renewal for imported certificates. For more information about the differences between certificates that you import and those that ACM provides, see Importing Certificates in the Amazon Web Services Certificate Manager User Guide.

certificateDetail_options :: Lens' CertificateDetail (Maybe CertificateOptions)

Value that specifies whether to add the certificate to a transparency log. Certificate transparency makes it possible to detect SSL certificates that have been mistakenly or maliciously issued. A browser might respond to certificate that has not been logged by showing an error message. The logs are cryptographically secure.

certificateDetail_issuedAt :: Lens' CertificateDetail (Maybe UTCTime)

The time at which the certificate was issued. This value exists only when the certificate type is AMAZON_ISSUED.

certificateDetail_signatureAlgorithm :: Lens' CertificateDetail (Maybe Text)

The algorithm that was used to sign the certificate.

certificateDetail_domainValidationOptions :: Lens' CertificateDetail (Maybe (NonEmpty DomainValidation))

Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED.

certificateDetail_issuer :: Lens' CertificateDetail (Maybe Text)

The name of the certificate authority that issued and signed the certificate.

certificateDetail_notAfter :: Lens' CertificateDetail (Maybe UTCTime)

The time after which the certificate is not valid.

certificateDetail_certificateAuthorityArn :: Lens' CertificateDetail (Maybe Text)

The Amazon Resource Name (ARN) of the ACM PCA private certificate authority (CA) that issued the certificate. This has the following format:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012