Network.AWS.CertificateManager.Types.DomainValidation

Description

Documentation

data DomainValidation

Contains information about the validation of each domain name in the certificate.

See: newDomainValidation smart constructor.

Constructors

DomainValidation'

Fields

resourceRecord :: Maybe ResourceRecord

Contains the CNAME record that you add to your DNS database for domain validation. For more information, see Use DNS to Validate Domain Ownership.

Note: The CNAME information that you need does not include the name of your domain. If you include your domain name in the DNS database CNAME record, validation fails. For example, if the name is "_a79865eb4cd1a6ab990a45779b4e0b96.yourdomain.com", only "_a79865eb4cd1a6ab990a45779b4e0b96" must be used.

validationMethod :: Maybe ValidationMethod

Specifies the domain validation method.

validationEmails :: Maybe [Text]

A list of email addresses that ACM used to send domain validation emails.

validationStatus :: Maybe DomainStatus

The validation status of the domain name. This can be one of the following values:

PENDING_VALIDATION

@@SUCCESS

@@FAILED

validationDomain :: Maybe Text

The domain name that ACM used to send domain validation emails.

domainName :: Text

A fully qualified domain name (FQDN) in the certificate. For example, www.example.com or example.com.

newDomainValidation

Arguments

:: Text
$sel:domainName:DomainValidation'

-> DomainValidation

Create a value of DomainValidation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceRecord:DomainValidation', domainValidation_resourceRecord - Contains the CNAME record that you add to your DNS database for domain validation. For more information, see Use DNS to Validate Domain Ownership.

Note: The CNAME information that you need does not include the name of your domain. If you include your domain name in the DNS database CNAME record, validation fails. For example, if the name is "_a79865eb4cd1a6ab990a45779b4e0b96.yourdomain.com", only "_a79865eb4cd1a6ab990a45779b4e0b96" must be used.

$sel:validationMethod:DomainValidation', domainValidation_validationMethod - Specifies the domain validation method.

$sel:validationEmails:DomainValidation', domainValidation_validationEmails - A list of email addresses that ACM used to send domain validation emails.

$sel:validationStatus:DomainValidation', domainValidation_validationStatus - The validation status of the domain name. This can be one of the following values:

PENDING_VALIDATION

@@SUCCESS

@@FAILED

$sel:validationDomain:DomainValidation', domainValidation_validationDomain - The domain name that ACM used to send domain validation emails.

$sel:domainName:DomainValidation', domainValidation_domainName - A fully qualified domain name (FQDN) in the certificate. For example, www.example.com or example.com.

domainValidation_resourceRecord :: Lens' DomainValidation (Maybe ResourceRecord)

Contains the CNAME record that you add to your DNS database for domain validation. For more information, see Use DNS to Validate Domain Ownership.

Note: The CNAME information that you need does not include the name of your domain. If you include your domain name in the DNS database CNAME record, validation fails. For example, if the name is "_a79865eb4cd1a6ab990a45779b4e0b96.yourdomain.com", only "_a79865eb4cd1a6ab990a45779b4e0b96" must be used.

domainValidation_validationMethod :: Lens' DomainValidation (Maybe ValidationMethod)

Specifies the domain validation method.

domainValidation_validationEmails :: Lens' DomainValidation (Maybe [Text])

A list of email addresses that ACM used to send domain validation emails.

domainValidation_validationStatus :: Lens' DomainValidation (Maybe DomainStatus)

The validation status of the domain name. This can be one of the following values:

PENDING_VALIDATION

@@SUCCESS

@@FAILED

domainValidation_validationDomain :: Lens' DomainValidation (Maybe Text)

The domain name that ACM used to send domain validation emails.

domainValidation_domainName :: Lens' DomainValidation Text

A fully qualified domain name (FQDN) in the certificate. For example, www.example.com or example.com.