Network.AWS.CertificateManager.Types.Filters

Description

Documentation

data Filters

This structure can be used in the ListCertificates action to filter the output of the certificate list.

See: newFilters smart constructor.

Constructors

Filters'

Fields

keyTypes :: Maybe [KeyAlgorithm]

Specify one or more algorithms that can be used to generate key pairs.

Default filtering returns only RSA_1024 and RSA_2048 certificates that have at least one domain. To return other certificate types, provide the desired type signatures in a comma-separated list. For example, "keyTypes": ["RSA_2048,RSA_4096"] returns both RSA_2048 and RSA_4096 certificates.

extendedKeyUsage :: Maybe [ExtendedKeyUsageName]

Specify one or more ExtendedKeyUsage extension values.

keyUsage :: Maybe [KeyUsageName]

Specify one or more KeyUsage extension values.

newFilters :: Filters

Create a value of Filters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyTypes:Filters', filters_keyTypes - Specify one or more algorithms that can be used to generate key pairs.

Default filtering returns only RSA_1024 and RSA_2048 certificates that have at least one domain. To return other certificate types, provide the desired type signatures in a comma-separated list. For example, "keyTypes": ["RSA_2048,RSA_4096"] returns both RSA_2048 and RSA_4096 certificates.

$sel:extendedKeyUsage:Filters', filters_extendedKeyUsage - Specify one or more ExtendedKeyUsage extension values.

$sel:keyUsage:Filters', filters_keyUsage - Specify one or more KeyUsage extension values.

filters_keyTypes :: Lens' Filters (Maybe [KeyAlgorithm])

Specify one or more algorithms that can be used to generate key pairs.

Default filtering returns only RSA_1024 and RSA_2048 certificates that have at least one domain. To return other certificate types, provide the desired type signatures in a comma-separated list. For example, "keyTypes": ["RSA_2048,RSA_4096"] returns both RSA_2048 and RSA_4096 certificates.

filters_extendedKeyUsage :: Lens' Filters (Maybe [ExtendedKeyUsageName])

Specify one or more ExtendedKeyUsage extension values.

filters_keyUsage :: Lens' Filters (Maybe [KeyUsageName])

Specify one or more KeyUsage extension values.