Network.AWS.CertificateManagerPCA.DeletePolicy

Description

Deletes the resource-based policy attached to a private CA. Deletion will remove any access that the policy has granted. If there is no policy attached to the private CA, this action will return successful.

If you delete a policy that was applied through AWS Resource Access Manager (RAM), the CA will be removed from all shares in which it was included.

The AWS Certificate Manager Service Linked Role that the policy supports is not affected when you delete the policy.

The current policy can be shown with GetPolicy and updated with PutPolicy.

About Policies

A policy grants access on a private CA to an AWS customer account, to AWS Organizations, or to an AWS Organizations unit. Policies are under the control of a CA administrator. For more information, see Using a Resource Based Policy with ACM Private CA.

A policy permits a user of AWS Certificate Manager (ACM) to issue ACM certificates signed by a CA in another account.

For ACM to manage automatic renewal of these certificates, the ACM user must configure a Service Linked Role (SLR). The SLR allows the ACM service to assume the identity of the user, subject to confirmation against the ACM Private CA policy. For more information, see Using a Service Linked Role with ACM.

Updates made in AWS Resource Manager (RAM) are reflected in policies. For more information, see Attach a Policy for Cross-Account Access.

Creating a Request

data DeletePolicy

See: newDeletePolicy smart constructor.

Constructors

DeletePolicy'

Fields

resourceArn :: Text

The Amazon Resource Number (ARN) of the private CA that will have its policy deleted. You can find the CA's ARN by calling the ListCertificateAuthorities action. The ARN value must have the form arn:aws:acm-pca:region:account:certificate-authority/01234567-89ab-cdef-0123-0123456789ab.

newDeletePolicy

Arguments

:: Text
$sel:resourceArn:DeletePolicy'

-> DeletePolicy

Create a value of DeletePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:DeletePolicy', deletePolicy_resourceArn - The Amazon Resource Number (ARN) of the private CA that will have its policy deleted. You can find the CA's ARN by calling the ListCertificateAuthorities action. The ARN value must have the form arn:aws:acm-pca:region:account:certificate-authority/01234567-89ab-cdef-0123-0123456789ab.

Request Lenses

deletePolicy_resourceArn :: Lens' DeletePolicy Text

The Amazon Resource Number (ARN) of the private CA that will have its policy deleted. You can find the CA's ARN by calling the ListCertificateAuthorities action. The ARN value must have the form arn:aws:acm-pca:region:account:certificate-authority/01234567-89ab-cdef-0123-0123456789ab.

Destructuring the Response

data DeletePolicyResponse

See: newDeletePolicyResponse smart constructor.

Constructors

DeletePolicyResponse'

newDeletePolicyResponse :: DeletePolicyResponse

Create a value of DeletePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.