Network.AWS.CertificateManagerPCA.DescribeCertificateAuthority

Description

Lists information about your private certificate authority (CA) or one that has been shared with you. You specify the private CA on input by its ARN (Amazon Resource Name). The output contains the status of your CA. This can be any of the following:

CREATING - ACM Private CA is creating your private certificate authority.

PENDING_CERTIFICATE - The certificate is pending. You must use your ACM Private CA-hosted or on-premises root or subordinate CA to sign your private CA CSR and then import it into PCA.

ACTIVE - Your private CA is active.

DISABLED - Your private CA has been disabled.

EXPIRED - Your private CA certificate has expired.

FAILED - Your private CA has failed. Your CA can fail because of problems such a network outage or back-end AWS failure or other errors. A failed CA can never return to the pending state. You must create a new CA.

DELETED - Your private CA is within the restoration period, after which it is permanently deleted. The length of time remaining in the CA's restoration period is also included in this action's output.

Creating a Request

data DescribeCertificateAuthority

See: newDescribeCertificateAuthority smart constructor.

Constructors

DescribeCertificateAuthority'

Fields

certificateAuthorityArn :: Text

The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

newDescribeCertificateAuthority

Arguments

:: Text
$sel:certificateAuthorityArn:DescribeCertificateAuthority'

-> DescribeCertificateAuthority

Create a value of DescribeCertificateAuthority with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:DescribeCertificateAuthority', describeCertificateAuthority_certificateAuthorityArn - The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

Request Lenses

describeCertificateAuthority_certificateAuthorityArn :: Lens' DescribeCertificateAuthority Text

The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

Destructuring the Response

data DescribeCertificateAuthorityResponse

See: newDescribeCertificateAuthorityResponse smart constructor.

Constructors

DescribeCertificateAuthorityResponse'

Fields

certificateAuthority :: Maybe CertificateAuthority

A CertificateAuthority structure that contains information about your private CA.

httpStatus :: Int

The response's http status code.

newDescribeCertificateAuthorityResponse

Arguments

:: Int
$sel:httpStatus:DescribeCertificateAuthorityResponse'

-> DescribeCertificateAuthorityResponse

Create a value of DescribeCertificateAuthorityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthority:DescribeCertificateAuthorityResponse', describeCertificateAuthorityResponse_certificateAuthority - A CertificateAuthority structure that contains information about your private CA.

$sel:httpStatus:DescribeCertificateAuthorityResponse', describeCertificateAuthorityResponse_httpStatus - The response's http status code.

Response Lenses

describeCertificateAuthorityResponse_certificateAuthority :: Lens' DescribeCertificateAuthorityResponse (Maybe CertificateAuthority)

A CertificateAuthority structure that contains information about your private CA.

describeCertificateAuthorityResponse_httpStatus :: Lens' DescribeCertificateAuthorityResponse Int

The response's http status code.