Network.AWS.CertificateManagerPCA.GetCertificate

Description

Retrieves a certificate from your private CA or one that has been shared with you. The ARN of the certificate is returned when you call the IssueCertificate action. You must specify both the ARN of your private CA and the ARN of the issued certificate when calling the GetCertificate action. You can retrieve the certificate if it is in the ISSUED state. You can call the CreateCertificateAuthorityAuditReport action to create a report that contains information about all of the certificates issued and revoked by your private CA.

Creating a Request

data GetCertificate

See: newGetCertificate smart constructor.

Constructors

GetCertificate'

Fields

certificateAuthorityArn :: Text

The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

certificateArn :: Text

The ARN of the issued certificate. The ARN contains the certificate serial number and must be in the following form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012/certificate/286535153982981100925020015808220737245

newGetCertificate

Arguments

:: Text
$sel:certificateAuthorityArn:GetCertificate'

-> Text
$sel:certificateArn:GetCertificate'

-> GetCertificate

Create a value of GetCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:GetCertificate', getCertificate_certificateAuthorityArn - The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

$sel:certificateArn:GetCertificate', getCertificate_certificateArn - The ARN of the issued certificate. The ARN contains the certificate serial number and must be in the following form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012/certificate/286535153982981100925020015808220737245

Request Lenses

getCertificate_certificateAuthorityArn :: Lens' GetCertificate Text

The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

getCertificate_certificateArn :: Lens' GetCertificate Text

The ARN of the issued certificate. The ARN contains the certificate serial number and must be in the following form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012/certificate/286535153982981100925020015808220737245

Destructuring the Response

data GetCertificateResponse

See: newGetCertificateResponse smart constructor.

Constructors

GetCertificateResponse'

Fields

certificateChain :: Maybe Text

The base64 PEM-encoded certificate chain that chains up to the root CA certificate that you used to sign your private CA certificate.

certificate :: Maybe Text

The base64 PEM-encoded certificate specified by the CertificateArn parameter.

httpStatus :: Int

The response's http status code.

newGetCertificateResponse

Arguments

:: Int
$sel:httpStatus:GetCertificateResponse'

-> GetCertificateResponse

Create a value of GetCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateChain:GetCertificateResponse', getCertificateResponse_certificateChain - The base64 PEM-encoded certificate chain that chains up to the root CA certificate that you used to sign your private CA certificate.

$sel:certificate:GetCertificateResponse', getCertificateResponse_certificate - The base64 PEM-encoded certificate specified by the CertificateArn parameter.

$sel:httpStatus:GetCertificateResponse', getCertificateResponse_httpStatus - The response's http status code.

Response Lenses

getCertificateResponse_certificateChain :: Lens' GetCertificateResponse (Maybe Text)

The base64 PEM-encoded certificate chain that chains up to the root CA certificate that you used to sign your private CA certificate.

getCertificateResponse_certificate :: Lens' GetCertificateResponse (Maybe Text)

The base64 PEM-encoded certificate specified by the CertificateArn parameter.

getCertificateResponse_httpStatus :: Lens' GetCertificateResponse Int

The response's http status code.