Network.AWS.CertificateManagerPCA.GetCertificateAuthorityCertificate

Description

Retrieves the certificate and certificate chain for your private certificate authority (CA) or one that has been shared with you. Both the certificate and the chain are base64 PEM-encoded. The chain does not include the CA certificate. Each certificate in the chain signs the one before it.

Creating a Request

data GetCertificateAuthorityCertificate

See: newGetCertificateAuthorityCertificate smart constructor.

Constructors

GetCertificateAuthorityCertificate'

Fields

certificateAuthorityArn :: Text

The Amazon Resource Name (ARN) of your private CA. This is of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

newGetCertificateAuthorityCertificate

Arguments

:: Text
$sel:certificateAuthorityArn:GetCertificateAuthorityCertificate'

-> GetCertificateAuthorityCertificate

Create a value of GetCertificateAuthorityCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:GetCertificateAuthorityCertificate', getCertificateAuthorityCertificate_certificateAuthorityArn - The Amazon Resource Name (ARN) of your private CA. This is of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

Request Lenses

getCertificateAuthorityCertificate_certificateAuthorityArn :: Lens' GetCertificateAuthorityCertificate Text

The Amazon Resource Name (ARN) of your private CA. This is of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 .

Destructuring the Response

data GetCertificateAuthorityCertificateResponse

See: newGetCertificateAuthorityCertificateResponse smart constructor.

Constructors

GetCertificateAuthorityCertificateResponse'

Fields

certificateChain :: Maybe Text

Base64-encoded certificate chain that includes any intermediate certificates and chains up to root certificate that you used to sign your private CA certificate. The chain does not include your private CA certificate. If this is a root CA, the value will be null.

certificate :: Maybe Text

Base64-encoded certificate authority (CA) certificate.

httpStatus :: Int

The response's http status code.

newGetCertificateAuthorityCertificateResponse

Arguments

:: Int
$sel:httpStatus:GetCertificateAuthorityCertificateResponse'

-> GetCertificateAuthorityCertificateResponse

Create a value of GetCertificateAuthorityCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateChain:GetCertificateAuthorityCertificateResponse', getCertificateAuthorityCertificateResponse_certificateChain - Base64-encoded certificate chain that includes any intermediate certificates and chains up to root certificate that you used to sign your private CA certificate. The chain does not include your private CA certificate. If this is a root CA, the value will be null.

$sel:certificate:GetCertificateAuthorityCertificateResponse', getCertificateAuthorityCertificateResponse_certificate - Base64-encoded certificate authority (CA) certificate.

$sel:httpStatus:GetCertificateAuthorityCertificateResponse', getCertificateAuthorityCertificateResponse_httpStatus - The response's http status code.

Response Lenses

getCertificateAuthorityCertificateResponse_certificateChain :: Lens' GetCertificateAuthorityCertificateResponse (Maybe Text)

Base64-encoded certificate chain that includes any intermediate certificates and chains up to root certificate that you used to sign your private CA certificate. The chain does not include your private CA certificate. If this is a root CA, the value will be null.

getCertificateAuthorityCertificateResponse_certificate :: Lens' GetCertificateAuthorityCertificateResponse (Maybe Text)

Base64-encoded certificate authority (CA) certificate.

getCertificateAuthorityCertificateResponse_httpStatus :: Lens' GetCertificateAuthorityCertificateResponse Int

The response's http status code.