Network.AWS.CertificateManagerPCA.ImportCertificateAuthorityCertificate

Description

Imports a signed private CA certificate into ACM Private CA. This action is used when you are using a chain of trust whose root is located outside ACM Private CA. Before you can call this action, the following preparations must in place:

In ACM Private CA, call the CreateCertificateAuthority action to create the private CA that you plan to back with the imported certificate.

Call the GetCertificateAuthorityCsr action to generate a certificate signing request (CSR).

Sign the CSR using a root or intermediate CA hosted by either an on-premises PKI hierarchy or by a commercial CA.

Create a certificate chain and copy the signed certificate and the certificate chain to your working directory.

ACM Private CA supports three scenarios for installing a CA certificate:

Installing a certificate for a root CA hosted by ACM Private CA.

Installing a subordinate CA certificate whose parent authority is hosted by ACM Private CA.

Installing a subordinate CA certificate whose parent authority is externally hosted.

The following additional requirements apply when you import a CA certificate.

Only a self-signed certificate can be imported as a root CA.

A self-signed certificate cannot be imported as a subordinate CA.

Your certificate chain must not include the private CA certificate that you are importing.

Your root CA must be the last certificate in your chain. The subordinate certificate, if any, that your root CA signed must be next to last. The subordinate certificate signed by the preceding subordinate CA must come next, and so on until your chain is built.

The chain must be PEM-encoded.

The maximum allowed size of a certificate is 32 KB.

The maximum allowed size of a certificate chain is 2 MB.

Enforcement of Critical Constraints

ACM Private CA allows the following extensions to be marked critical in the imported CA certificate or chain.

Basic constraints (must be marked critical)

Subject alternative names

Key usage

Extended key usage

Authority key identifier

Subject key identifier

Issuer alternative name

Subject directory attributes

Subject information access

Certificate policies

Policy mappings

Inhibit anyPolicy

ACM Private CA rejects the following extensions when they are marked critical in an imported CA certificate or chain.

Name constraints

Policy constraints

CRL distribution points

Authority information access

Freshest CRL

Any other extension

Creating a Request

data ImportCertificateAuthorityCertificate

See: newImportCertificateAuthorityCertificate smart constructor.

Constructors

ImportCertificateAuthorityCertificate'

Fields

certificateChain :: Maybe Base64

A PEM-encoded file that contains all of your certificates, other than the certificate you're importing, chaining up to your root CA. Your ACM Private CA-hosted or on-premises root certificate is the last in the chain, and each certificate in the chain signs the one preceding.

This parameter must be supplied when you import a subordinate CA. When you import a root CA, there is no chain.

certificateAuthorityArn :: Text

The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

certificate :: Base64

The PEM-encoded certificate for a private CA. This may be a self-signed certificate in the case of a root CA, or it may be signed by another CA that you control.

newImportCertificateAuthorityCertificate

Arguments

:: Text
$sel:certificateAuthorityArn:ImportCertificateAuthorityCertificate'

-> ByteString
$sel:certificate:ImportCertificateAuthorityCertificate'

-> ImportCertificateAuthorityCertificate

Create a value of ImportCertificateAuthorityCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateChain:ImportCertificateAuthorityCertificate', importCertificateAuthorityCertificate_certificateChain - A PEM-encoded file that contains all of your certificates, other than the certificate you're importing, chaining up to your root CA. Your ACM Private CA-hosted or on-premises root certificate is the last in the chain, and each certificate in the chain signs the one preceding.

This parameter must be supplied when you import a subordinate CA. When you import a root CA, there is no chain.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:certificateAuthorityArn:ImportCertificateAuthorityCertificate', importCertificateAuthorityCertificate_certificateAuthorityArn - The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

$sel:certificate:ImportCertificateAuthorityCertificate', importCertificateAuthorityCertificate_certificate - The PEM-encoded certificate for a private CA. This may be a self-signed certificate in the case of a root CA, or it may be signed by another CA that you control.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

Request Lenses

importCertificateAuthorityCertificate_certificateChain :: Lens' ImportCertificateAuthorityCertificate (Maybe ByteString)

A PEM-encoded file that contains all of your certificates, other than the certificate you're importing, chaining up to your root CA. Your ACM Private CA-hosted or on-premises root certificate is the last in the chain, and each certificate in the chain signs the one preceding.

This parameter must be supplied when you import a subordinate CA. When you import a root CA, there is no chain.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

importCertificateAuthorityCertificate_certificateAuthorityArn :: Lens' ImportCertificateAuthorityCertificate Text

The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

importCertificateAuthorityCertificate_certificate :: Lens' ImportCertificateAuthorityCertificate ByteString

The PEM-encoded certificate for a private CA. This may be a self-signed certificate in the case of a root CA, or it may be signed by another CA that you control.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

Destructuring the Response

data ImportCertificateAuthorityCertificateResponse

See: newImportCertificateAuthorityCertificateResponse smart constructor.

Constructors

ImportCertificateAuthorityCertificateResponse'

newImportCertificateAuthorityCertificateResponse :: ImportCertificateAuthorityCertificateResponse

Create a value of ImportCertificateAuthorityCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.