Network.AWS.CertificateManagerPCA.ListPermissions

Description

List all permissions on a private CA, if any, granted to the AWS Certificate Manager (ACM) service principal (acm.amazonaws.com).

These permissions allow ACM to issue and renew ACM certificates that reside in the same AWS account as the CA.

Permissions can be granted with the CreatePermission action and revoked with the DeletePermission action.

About Permissions

If the private CA and the certificates it issues reside in the same account, you can use CreatePermission to grant permissions for ACM to carry out automatic certificate renewals.

For automatic certificate renewal to succeed, the ACM service principal needs permissions to create, retrieve, and list certificates.

If the private CA and the ACM certificates reside in different accounts, then permissions cannot be used to enable automatic renewals. Instead, the ACM certificate owner must set up a resource-based policy to enable cross-account issuance and renewals. For more information, see Using a Resource Based Policy with ACM Private CA.

This operation returns paginated results.

Creating a Request

data ListPermissions

See: newListPermissions smart constructor.

Constructors

ListPermissions'

Fields

nextToken :: Maybe Text

When paginating results, use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextToken from the response you just received.

maxResults :: Maybe Natural

When paginating results, use this parameter to specify the maximum number of items to return in the response. If additional items exist beyond the number you specify, the NextToken element is sent in the response. Use this NextToken value in a subsequent request to retrieve additional items.

certificateAuthorityArn :: Text

The Amazon Resource Number (ARN) of the private CA to inspect. You can find the ARN by calling the ListCertificateAuthorities action. This must be of the form: arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 You can get a private CA's ARN by running the ListCertificateAuthorities action.

newListPermissions

Arguments

:: Text
$sel:certificateAuthorityArn:ListPermissions'

-> ListPermissions

Create a value of ListPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPermissions', listPermissions_nextToken - When paginating results, use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextToken from the response you just received.

$sel:maxResults:ListPermissions', listPermissions_maxResults - When paginating results, use this parameter to specify the maximum number of items to return in the response. If additional items exist beyond the number you specify, the NextToken element is sent in the response. Use this NextToken value in a subsequent request to retrieve additional items.

$sel:certificateAuthorityArn:ListPermissions', listPermissions_certificateAuthorityArn - The Amazon Resource Number (ARN) of the private CA to inspect. You can find the ARN by calling the ListCertificateAuthorities action. This must be of the form: arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 You can get a private CA's ARN by running the ListCertificateAuthorities action.

Request Lenses

listPermissions_nextToken :: Lens' ListPermissions (Maybe Text)

When paginating results, use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextToken from the response you just received.

listPermissions_maxResults :: Lens' ListPermissions (Maybe Natural)

When paginating results, use this parameter to specify the maximum number of items to return in the response. If additional items exist beyond the number you specify, the NextToken element is sent in the response. Use this NextToken value in a subsequent request to retrieve additional items.

listPermissions_certificateAuthorityArn :: Lens' ListPermissions Text

The Amazon Resource Number (ARN) of the private CA to inspect. You can find the ARN by calling the ListCertificateAuthorities action. This must be of the form: arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 You can get a private CA's ARN by running the ListCertificateAuthorities action.

Destructuring the Response

data ListPermissionsResponse

See: newListPermissionsResponse smart constructor.

Constructors

ListPermissionsResponse'

Fields

nextToken :: Maybe Text

When the list is truncated, this value is present and should be used for the NextToken parameter in a subsequent pagination request.

permissions :: Maybe [Permission]

Summary information about each permission assigned by the specified private CA, including the action enabled, the policy provided, and the time of creation.

httpStatus :: Int

The response's http status code.

newListPermissionsResponse

Arguments

:: Int
$sel:httpStatus:ListPermissionsResponse'

-> ListPermissionsResponse

Create a value of ListPermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPermissions', listPermissionsResponse_nextToken - When the list is truncated, this value is present and should be used for the NextToken parameter in a subsequent pagination request.

$sel:permissions:ListPermissionsResponse', listPermissionsResponse_permissions - Summary information about each permission assigned by the specified private CA, including the action enabled, the policy provided, and the time of creation.

$sel:httpStatus:ListPermissionsResponse', listPermissionsResponse_httpStatus - The response's http status code.

Response Lenses

listPermissionsResponse_nextToken :: Lens' ListPermissionsResponse (Maybe Text)

When the list is truncated, this value is present and should be used for the NextToken parameter in a subsequent pagination request.

listPermissionsResponse_permissions :: Lens' ListPermissionsResponse (Maybe [Permission])

Summary information about each permission assigned by the specified private CA, including the action enabled, the policy provided, and the time of creation.

listPermissionsResponse_httpStatus :: Lens' ListPermissionsResponse Int

The response's http status code.