Network.AWS.CertificateManagerPCA.PutPolicy

Description

Attaches a resource-based policy to a private CA.

A policy can also be applied by sharing a private CA through AWS Resource Access Manager (RAM). For more information, see Attach a Policy for Cross-Account Access.

The policy can be displayed with GetPolicy and removed with DeletePolicy.

About Policies

A policy grants access on a private CA to an AWS customer account, to AWS Organizations, or to an AWS Organizations unit. Policies are under the control of a CA administrator. For more information, see Using a Resource Based Policy with ACM Private CA.

A policy permits a user of AWS Certificate Manager (ACM) to issue ACM certificates signed by a CA in another account.

For ACM to manage automatic renewal of these certificates, the ACM user must configure a Service Linked Role (SLR). The SLR allows the ACM service to assume the identity of the user, subject to confirmation against the ACM Private CA policy. For more information, see Using a Service Linked Role with ACM.

Updates made in AWS Resource Manager (RAM) are reflected in policies. For more information, see Attach a Policy for Cross-Account Access.

Creating a Request

data PutPolicy

See: newPutPolicy smart constructor.

Constructors

PutPolicy'

Fields

resourceArn :: Text

The Amazon Resource Number (ARN) of the private CA to associate with the policy. The ARN of the CA can be found by calling the ListCertificateAuthorities action.

policy :: Text

The path and file name of a JSON-formatted IAM policy to attach to the specified private CA resource. If this policy does not contain all required statements or if it includes any statement that is not allowed, the PutPolicy action returns an InvalidPolicyException. For information about IAM policy and statement structure, see Overview of JSON Policies.

newPutPolicy

Arguments

:: Text
$sel:resourceArn:PutPolicy'

-> Text
$sel:policy:PutPolicy'

-> PutPolicy

Create a value of PutPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:PutPolicy', putPolicy_resourceArn - The Amazon Resource Number (ARN) of the private CA to associate with the policy. The ARN of the CA can be found by calling the ListCertificateAuthorities action.

$sel:policy:PutPolicy', putPolicy_policy - The path and file name of a JSON-formatted IAM policy to attach to the specified private CA resource. If this policy does not contain all required statements or if it includes any statement that is not allowed, the PutPolicy action returns an InvalidPolicyException. For information about IAM policy and statement structure, see Overview of JSON Policies.

Request Lenses

putPolicy_resourceArn :: Lens' PutPolicy Text

The Amazon Resource Number (ARN) of the private CA to associate with the policy. The ARN of the CA can be found by calling the ListCertificateAuthorities action.

putPolicy_policy :: Lens' PutPolicy Text

The path and file name of a JSON-formatted IAM policy to attach to the specified private CA resource. If this policy does not contain all required statements or if it includes any statement that is not allowed, the PutPolicy action returns an InvalidPolicyException. For information about IAM policy and statement structure, see Overview of JSON Policies.

Destructuring the Response

data PutPolicyResponse

See: newPutPolicyResponse smart constructor.

Constructors

PutPolicyResponse'

newPutPolicyResponse :: PutPolicyResponse

Create a value of PutPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.