Network.AWS.CertificateManagerPCA.RevokeCertificate

Description

Revokes a certificate that was issued inside ACM Private CA. If you enable a certificate revocation list (CRL) when you create or update your private CA, information about the revoked certificates will be included in the CRL. ACM Private CA writes the CRL to an S3 bucket that you specify. A CRL is typically updated approximately 30 minutes after a certificate is revoked. If for any reason the CRL update fails, ACM Private CA attempts makes further attempts every 15 minutes. With Amazon CloudWatch, you can create alarms for the metrics CRLGenerated and MisconfiguredCRLBucket. For more information, see Supported CloudWatch Metrics.

Both PCA and the IAM principal must have permission to write to the S3 bucket that you specify. If the IAM principal making the call does not have permission to write to the bucket, then an exception is thrown. For more information, see Configure Access to ACM Private CA.

ACM Private CA also writes revocation information to the audit report. For more information, see CreateCertificateAuthorityAuditReport.

You cannot revoke a root CA self-signed certificate.

Creating a Request

data RevokeCertificate

See: newRevokeCertificate smart constructor.

Constructors

RevokeCertificate'

Fields

certificateAuthorityArn :: Text

Amazon Resource Name (ARN) of the private CA that issued the certificate to be revoked. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

certificateSerial :: Text

Serial number of the certificate to be revoked. This must be in hexadecimal format. You can retrieve the serial number by calling GetCertificate with the Amazon Resource Name (ARN) of the certificate you want and the ARN of your private CA. The GetCertificate action retrieves the certificate in the PEM format. You can use the following OpenSSL command to list the certificate in text format and copy the hexadecimal serial number.

openssl x509 -in file_path -text -noout

You can also copy the serial number from the console or use the DescribeCertificate action in the AWS Certificate Manager API Reference.

revocationReason :: RevocationReason

Specifies why you revoked the certificate.

newRevokeCertificate

Arguments

:: Text
$sel:certificateAuthorityArn:RevokeCertificate'

-> Text
$sel:certificateSerial:RevokeCertificate'

-> RevocationReason
$sel:revocationReason:RevokeCertificate'

-> RevokeCertificate

Create a value of RevokeCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:RevokeCertificate', revokeCertificate_certificateAuthorityArn - Amazon Resource Name (ARN) of the private CA that issued the certificate to be revoked. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

$sel:certificateSerial:RevokeCertificate', revokeCertificate_certificateSerial - Serial number of the certificate to be revoked. This must be in hexadecimal format. You can retrieve the serial number by calling GetCertificate with the Amazon Resource Name (ARN) of the certificate you want and the ARN of your private CA. The GetCertificate action retrieves the certificate in the PEM format. You can use the following OpenSSL command to list the certificate in text format and copy the hexadecimal serial number.

openssl x509 -in file_path -text -noout

You can also copy the serial number from the console or use the DescribeCertificate action in the AWS Certificate Manager API Reference.

$sel:revocationReason:RevokeCertificate', revokeCertificate_revocationReason - Specifies why you revoked the certificate.

Request Lenses

revokeCertificate_certificateAuthorityArn :: Lens' RevokeCertificate Text

Amazon Resource Name (ARN) of the private CA that issued the certificate to be revoked. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

revokeCertificate_certificateSerial :: Lens' RevokeCertificate Text

Serial number of the certificate to be revoked. This must be in hexadecimal format. You can retrieve the serial number by calling GetCertificate with the Amazon Resource Name (ARN) of the certificate you want and the ARN of your private CA. The GetCertificate action retrieves the certificate in the PEM format. You can use the following OpenSSL command to list the certificate in text format and copy the hexadecimal serial number.

openssl x509 -in file_path -text -noout

You can also copy the serial number from the console or use the DescribeCertificate action in the AWS Certificate Manager API Reference.

revokeCertificate_revocationReason :: Lens' RevokeCertificate RevocationReason

Specifies why you revoked the certificate.

Destructuring the Response

data RevokeCertificateResponse

See: newRevokeCertificateResponse smart constructor.

Constructors

RevokeCertificateResponse'

newRevokeCertificateResponse :: RevokeCertificateResponse

Create a value of RevokeCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.