Network.AWS.CertificateManagerPCA.TagCertificateAuthority

Description

Adds one or more tags to your private CA. Tags are labels that you can use to identify and organize your AWS resources. Each tag consists of a key and an optional value. You specify the private CA on input by its Amazon Resource Name (ARN). You specify the tag by using a key-value pair. You can apply a tag to just one private CA if you want to identify a specific characteristic of that CA, or you can apply the same tag to multiple private CAs if you want to filter for a common relationship among those CAs. To remove one or more tags, use the UntagCertificateAuthority action. Call the ListTags action to see what tags are associated with your CA.

Creating a Request

data TagCertificateAuthority

See: newTagCertificateAuthority smart constructor.

Constructors

TagCertificateAuthority'

Fields

certificateAuthorityArn :: Text

The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

tags :: NonEmpty Tag

List of tags to be associated with the CA.

newTagCertificateAuthority

Arguments

:: Text
$sel:certificateAuthorityArn:TagCertificateAuthority'

-> NonEmpty Tag
$sel:tags:TagCertificateAuthority'

-> TagCertificateAuthority

Create a value of TagCertificateAuthority with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:TagCertificateAuthority', tagCertificateAuthority_certificateAuthorityArn - The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

$sel:tags:TagCertificateAuthority', tagCertificateAuthority_tags - List of tags to be associated with the CA.

Request Lenses

tagCertificateAuthority_certificateAuthorityArn :: Lens' TagCertificateAuthority Text

The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

tagCertificateAuthority_tags :: Lens' TagCertificateAuthority (NonEmpty Tag)

List of tags to be associated with the CA.

Destructuring the Response

data TagCertificateAuthorityResponse

See: newTagCertificateAuthorityResponse smart constructor.

Constructors

TagCertificateAuthorityResponse'

newTagCertificateAuthorityResponse :: TagCertificateAuthorityResponse

Create a value of TagCertificateAuthorityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.