Network.AWS.CertificateManagerPCA.Types.ASN1Subject

Description

Documentation

data ASN1Subject

Contains information about the certificate subject. The Subject field in the certificate identifies the entity that owns or controls the public key in the certificate. The entity can be a user, computer, device, or service. The Subject must contain an X.500 distinguished name (DN). A DN is a sequence of relative distinguished names (RDNs). The RDNs are separated by commas in the certificate.

See: newASN1Subject smart constructor.

Constructors

ASN1Subject'

Fields

locality :: Maybe Text

The locality (such as a city or town) in which the certificate subject is located.

generationQualifier :: Maybe Text

Typically a qualifier appended to the name of an individual. Examples include Jr. for junior, Sr. for senior, and III for third.

surname :: Maybe Text

Family name. In the US and the UK, for example, the surname of an individual is ordered last. In Asian cultures the surname is typically ordered first.

organizationalUnit :: Maybe Text

A subdivision or unit of the organization (such as sales or finance) with which the certificate subject is affiliated.

title :: Maybe Text

A title such as Mr. or Ms., which is pre-pended to the name to refer formally to the certificate subject.

initials :: Maybe Text

Concatenation that typically contains the first letter of the GivenName, the first letter of the middle name if one exists, and the first letter of the Surname.

commonName :: Maybe Text

For CA and end-entity certificates in a private PKI, the common name (CN) can be any string within the length limit.

Note: In publicly trusted certificates, the common name must be a fully qualified domain name (FQDN) associated with the certificate subject.

pseudonym :: Maybe Text

Typically a shortened version of a longer GivenName. For example, Jonathan is often shortened to John. Elizabeth is often shortened to Beth, Liz, or Eliza.

state :: Maybe Text

State in which the subject of the certificate is located.

givenName :: Maybe Text

First name.

serialNumber :: Maybe Text

The certificate serial number.

distinguishedNameQualifier :: Maybe Text

Disambiguating information for the certificate subject.

organization :: Maybe Text

Legal name of the organization with which the certificate subject is affiliated.

country :: Maybe Text

Two-digit code that specifies the country in which the certificate subject located.

newASN1Subject :: ASN1Subject

Create a value of ASN1Subject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locality:ASN1Subject', aSN1Subject_locality - The locality (such as a city or town) in which the certificate subject is located.

$sel:generationQualifier:ASN1Subject', aSN1Subject_generationQualifier - Typically a qualifier appended to the name of an individual. Examples include Jr. for junior, Sr. for senior, and III for third.

$sel:surname:ASN1Subject', aSN1Subject_surname - Family name. In the US and the UK, for example, the surname of an individual is ordered last. In Asian cultures the surname is typically ordered first.

$sel:organizationalUnit:ASN1Subject', aSN1Subject_organizationalUnit - A subdivision or unit of the organization (such as sales or finance) with which the certificate subject is affiliated.

$sel:title:ASN1Subject', aSN1Subject_title - A title such as Mr. or Ms., which is pre-pended to the name to refer formally to the certificate subject.

$sel:initials:ASN1Subject', aSN1Subject_initials - Concatenation that typically contains the first letter of the GivenName, the first letter of the middle name if one exists, and the first letter of the Surname.

$sel:commonName:ASN1Subject', aSN1Subject_commonName - For CA and end-entity certificates in a private PKI, the common name (CN) can be any string within the length limit.

Note: In publicly trusted certificates, the common name must be a fully qualified domain name (FQDN) associated with the certificate subject.

$sel:pseudonym:ASN1Subject', aSN1Subject_pseudonym - Typically a shortened version of a longer GivenName. For example, Jonathan is often shortened to John. Elizabeth is often shortened to Beth, Liz, or Eliza.

$sel:state:ASN1Subject', aSN1Subject_state - State in which the subject of the certificate is located.

$sel:givenName:ASN1Subject', aSN1Subject_givenName - First name.

$sel:serialNumber:ASN1Subject', aSN1Subject_serialNumber - The certificate serial number.

$sel:distinguishedNameQualifier:ASN1Subject', aSN1Subject_distinguishedNameQualifier - Disambiguating information for the certificate subject.

$sel:organization:ASN1Subject', aSN1Subject_organization - Legal name of the organization with which the certificate subject is affiliated.

$sel:country:ASN1Subject', aSN1Subject_country - Two-digit code that specifies the country in which the certificate subject located.

aSN1Subject_locality :: Lens' ASN1Subject (Maybe Text)

The locality (such as a city or town) in which the certificate subject is located.

aSN1Subject_generationQualifier :: Lens' ASN1Subject (Maybe Text)

Typically a qualifier appended to the name of an individual. Examples include Jr. for junior, Sr. for senior, and III for third.

aSN1Subject_surname :: Lens' ASN1Subject (Maybe Text)

Family name. In the US and the UK, for example, the surname of an individual is ordered last. In Asian cultures the surname is typically ordered first.

aSN1Subject_organizationalUnit :: Lens' ASN1Subject (Maybe Text)

A subdivision or unit of the organization (such as sales or finance) with which the certificate subject is affiliated.

aSN1Subject_title :: Lens' ASN1Subject (Maybe Text)

A title such as Mr. or Ms., which is pre-pended to the name to refer formally to the certificate subject.

aSN1Subject_initials :: Lens' ASN1Subject (Maybe Text)

Concatenation that typically contains the first letter of the GivenName, the first letter of the middle name if one exists, and the first letter of the Surname.

aSN1Subject_commonName :: Lens' ASN1Subject (Maybe Text)

For CA and end-entity certificates in a private PKI, the common name (CN) can be any string within the length limit.

Note: In publicly trusted certificates, the common name must be a fully qualified domain name (FQDN) associated with the certificate subject.

aSN1Subject_pseudonym :: Lens' ASN1Subject (Maybe Text)

Typically a shortened version of a longer GivenName. For example, Jonathan is often shortened to John. Elizabeth is often shortened to Beth, Liz, or Eliza.

aSN1Subject_state :: Lens' ASN1Subject (Maybe Text)

State in which the subject of the certificate is located.

aSN1Subject_givenName :: Lens' ASN1Subject (Maybe Text)

First name.

aSN1Subject_serialNumber :: Lens' ASN1Subject (Maybe Text)

The certificate serial number.

aSN1Subject_distinguishedNameQualifier :: Lens' ASN1Subject (Maybe Text)

Disambiguating information for the certificate subject.

aSN1Subject_organization :: Lens' ASN1Subject (Maybe Text)

Legal name of the organization with which the certificate subject is affiliated.

aSN1Subject_country :: Lens' ASN1Subject (Maybe Text)

Two-digit code that specifies the country in which the certificate subject located.