Network.AWS.CertificateManagerPCA.Types.CertificateAuthority

Description

Documentation

data CertificateAuthority

Contains information about your private certificate authority (CA). Your private CA can issue and revoke X.509 digital certificates. Digital certificates verify that the entity named in the certificate Subject field owns or controls the public key contained in the __Subject Public Key Info__ field. Call the CreateCertificateAuthority action to create your private CA. You must then call the GetCertificateAuthorityCertificate action to retrieve a private CA certificate signing request (CSR). Sign the CSR with your ACM Private CA-hosted or on-premises root or subordinate CA certificate. Call the ImportCertificateAuthorityCertificate action to import the signed certificate into AWS Certificate Manager (ACM).

See: newCertificateAuthority smart constructor.

Constructors

CertificateAuthority'

Fields

status :: Maybe CertificateAuthorityStatus

Status of your private CA.

notBefore :: Maybe POSIX

Date and time before which your private CA certificate is not valid.

keyStorageSecurityStandard :: Maybe KeyStorageSecurityStandard

Defines a cryptographic key management compliance standard used for handling CA keys.

Default: FIPS_140_2_LEVEL_3_OR_HIGHER

Note: AWS Region ap-northeast-3 supports only FIPS_140_2_LEVEL_2_OR_HIGHER. You must explicitly specify this parameter and value when creating a CA in that Region. Specifying a different value (or no value) results in an InvalidArgsException with the message "A certificate authority cannot be created in this region with the specified security standard."

revocationConfiguration :: Maybe RevocationConfiguration

Information about the Online Certificate Status Protocol (OCSP) configuration or certificate revocation list (CRL) created and maintained by your private CA.

serial :: Maybe Text

Serial number of your private CA.

createdAt :: Maybe POSIX

Date and time at which your private CA was created.

arn :: Maybe Text

Amazon Resource Name (ARN) for your private certificate authority (CA). The format is 12345678-1234-1234-1234-123456789012 .

certificateAuthorityConfiguration :: Maybe CertificateAuthorityConfiguration

Your private CA configuration.

failureReason :: Maybe FailureReason

Reason the request to create your private CA failed.

notAfter :: Maybe POSIX

Date and time after which your private CA certificate is not valid.

lastStateChangeAt :: Maybe POSIX

Date and time at which your private CA was last updated.

type' :: Maybe CertificateAuthorityType

Type of your private CA.

ownerAccount :: Maybe Text

The AWS account ID that owns the certificate authority.

restorableUntil :: Maybe POSIX

The period during which a deleted CA can be restored. For more information, see the PermanentDeletionTimeInDays parameter of the DeleteCertificateAuthorityRequest action.

newCertificateAuthority :: CertificateAuthority

Create a value of CertificateAuthority with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CertificateAuthority', certificateAuthority_status - Status of your private CA.

$sel:notBefore:CertificateAuthority', certificateAuthority_notBefore - Date and time before which your private CA certificate is not valid.

$sel:keyStorageSecurityStandard:CertificateAuthority', certificateAuthority_keyStorageSecurityStandard - Defines a cryptographic key management compliance standard used for handling CA keys.

Default: FIPS_140_2_LEVEL_3_OR_HIGHER

Note: AWS Region ap-northeast-3 supports only FIPS_140_2_LEVEL_2_OR_HIGHER. You must explicitly specify this parameter and value when creating a CA in that Region. Specifying a different value (or no value) results in an InvalidArgsException with the message "A certificate authority cannot be created in this region with the specified security standard."

$sel:revocationConfiguration:CertificateAuthority', certificateAuthority_revocationConfiguration - Information about the Online Certificate Status Protocol (OCSP) configuration or certificate revocation list (CRL) created and maintained by your private CA.

$sel:serial:CertificateAuthority', certificateAuthority_serial - Serial number of your private CA.

$sel:createdAt:CertificateAuthority', certificateAuthority_createdAt - Date and time at which your private CA was created.

$sel:arn:CertificateAuthority', certificateAuthority_arn - Amazon Resource Name (ARN) for your private certificate authority (CA). The format is 12345678-1234-1234-1234-123456789012 .

$sel:certificateAuthorityConfiguration:CertificateAuthority', certificateAuthority_certificateAuthorityConfiguration - Your private CA configuration.

$sel:failureReason:CertificateAuthority', certificateAuthority_failureReason - Reason the request to create your private CA failed.

$sel:notAfter:CertificateAuthority', certificateAuthority_notAfter - Date and time after which your private CA certificate is not valid.

$sel:lastStateChangeAt:CertificateAuthority', certificateAuthority_lastStateChangeAt - Date and time at which your private CA was last updated.

$sel:type':CertificateAuthority', certificateAuthority_type - Type of your private CA.

$sel:ownerAccount:CertificateAuthority', certificateAuthority_ownerAccount - The AWS account ID that owns the certificate authority.

$sel:restorableUntil:CertificateAuthority', certificateAuthority_restorableUntil - The period during which a deleted CA can be restored. For more information, see the PermanentDeletionTimeInDays parameter of the DeleteCertificateAuthorityRequest action.

certificateAuthority_status :: Lens' CertificateAuthority (Maybe CertificateAuthorityStatus)

Status of your private CA.

certificateAuthority_notBefore :: Lens' CertificateAuthority (Maybe UTCTime)

Date and time before which your private CA certificate is not valid.

certificateAuthority_keyStorageSecurityStandard :: Lens' CertificateAuthority (Maybe KeyStorageSecurityStandard)

Defines a cryptographic key management compliance standard used for handling CA keys.

Default: FIPS_140_2_LEVEL_3_OR_HIGHER

Note: AWS Region ap-northeast-3 supports only FIPS_140_2_LEVEL_2_OR_HIGHER. You must explicitly specify this parameter and value when creating a CA in that Region. Specifying a different value (or no value) results in an InvalidArgsException with the message "A certificate authority cannot be created in this region with the specified security standard."

certificateAuthority_revocationConfiguration :: Lens' CertificateAuthority (Maybe RevocationConfiguration)

Information about the Online Certificate Status Protocol (OCSP) configuration or certificate revocation list (CRL) created and maintained by your private CA.

certificateAuthority_serial :: Lens' CertificateAuthority (Maybe Text)

Serial number of your private CA.

certificateAuthority_createdAt :: Lens' CertificateAuthority (Maybe UTCTime)

Date and time at which your private CA was created.

certificateAuthority_arn :: Lens' CertificateAuthority (Maybe Text)

Amazon Resource Name (ARN) for your private certificate authority (CA). The format is 12345678-1234-1234-1234-123456789012 .

certificateAuthority_certificateAuthorityConfiguration :: Lens' CertificateAuthority (Maybe CertificateAuthorityConfiguration)

Your private CA configuration.

certificateAuthority_failureReason :: Lens' CertificateAuthority (Maybe FailureReason)

Reason the request to create your private CA failed.

certificateAuthority_notAfter :: Lens' CertificateAuthority (Maybe UTCTime)

Date and time after which your private CA certificate is not valid.

certificateAuthority_lastStateChangeAt :: Lens' CertificateAuthority (Maybe UTCTime)

Date and time at which your private CA was last updated.

certificateAuthority_type :: Lens' CertificateAuthority (Maybe CertificateAuthorityType)

Type of your private CA.

certificateAuthority_ownerAccount :: Lens' CertificateAuthority (Maybe Text)

The AWS account ID that owns the certificate authority.

certificateAuthority_restorableUntil :: Lens' CertificateAuthority (Maybe UTCTime)

The period during which a deleted CA can be restored. For more information, see the PermanentDeletionTimeInDays parameter of the DeleteCertificateAuthorityRequest action.