Network.AWS.CertificateManagerPCA.Types.CertificateAuthorityConfiguration

Description

Documentation

data CertificateAuthorityConfiguration

Contains configuration information for your private certificate authority (CA). This includes information about the class of public key algorithm and the key pair that your private CA creates when it issues a certificate. It also includes the signature algorithm that it uses when issuing certificates, and its X.500 distinguished name. You must specify this information when you call the CreateCertificateAuthority action.

See: newCertificateAuthorityConfiguration smart constructor.

Constructors

CertificateAuthorityConfiguration'

Fields

csrExtensions :: Maybe CsrExtensions

Specifies information to be added to the extension section of the certificate signing request (CSR).

keyAlgorithm :: KeyAlgorithm

Type of the public key algorithm and size, in bits, of the key pair that your CA creates when it issues a certificate. When you create a subordinate CA, you must use a key algorithm supported by the parent CA.

signingAlgorithm :: SigningAlgorithm

Name of the algorithm your private CA uses to sign certificate requests.

This parameter should not be confused with the SigningAlgorithm parameter used to sign certificates when they are issued.

subject :: ASN1Subject

Structure that contains X.500 distinguished name information for your private CA.

newCertificateAuthorityConfiguration

Arguments

:: KeyAlgorithm
$sel:keyAlgorithm:CertificateAuthorityConfiguration'

-> SigningAlgorithm
$sel:signingAlgorithm:CertificateAuthorityConfiguration'

-> ASN1Subject
$sel:subject:CertificateAuthorityConfiguration'

-> CertificateAuthorityConfiguration

Create a value of CertificateAuthorityConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:csrExtensions:CertificateAuthorityConfiguration', certificateAuthorityConfiguration_csrExtensions - Specifies information to be added to the extension section of the certificate signing request (CSR).

$sel:keyAlgorithm:CertificateAuthorityConfiguration', certificateAuthorityConfiguration_keyAlgorithm - Type of the public key algorithm and size, in bits, of the key pair that your CA creates when it issues a certificate. When you create a subordinate CA, you must use a key algorithm supported by the parent CA.

$sel:signingAlgorithm:CertificateAuthorityConfiguration', certificateAuthorityConfiguration_signingAlgorithm - Name of the algorithm your private CA uses to sign certificate requests.

This parameter should not be confused with the SigningAlgorithm parameter used to sign certificates when they are issued.

$sel:subject:CertificateAuthorityConfiguration', certificateAuthorityConfiguration_subject - Structure that contains X.500 distinguished name information for your private CA.

certificateAuthorityConfiguration_csrExtensions :: Lens' CertificateAuthorityConfiguration (Maybe CsrExtensions)

Specifies information to be added to the extension section of the certificate signing request (CSR).

certificateAuthorityConfiguration_keyAlgorithm :: Lens' CertificateAuthorityConfiguration KeyAlgorithm

Type of the public key algorithm and size, in bits, of the key pair that your CA creates when it issues a certificate. When you create a subordinate CA, you must use a key algorithm supported by the parent CA.

certificateAuthorityConfiguration_signingAlgorithm :: Lens' CertificateAuthorityConfiguration SigningAlgorithm

Name of the algorithm your private CA uses to sign certificate requests.

This parameter should not be confused with the SigningAlgorithm parameter used to sign certificates when they are issued.

certificateAuthorityConfiguration_subject :: Lens' CertificateAuthorityConfiguration ASN1Subject

Structure that contains X.500 distinguished name information for your private CA.