Network.AWS.CertificateManagerPCA.Types.KeyUsage

Description

Documentation

data KeyUsage

Defines one or more purposes for which the key contained in the certificate can be used. Default value for each option is false.

See: newKeyUsage smart constructor.

Constructors

KeyUsage'

Fields

dataEncipherment :: Maybe Bool

Key can be used to decipher data.

keyCertSign :: Maybe Bool

Key can be used to sign certificates.

cRLSign :: Maybe Bool

Key can be used to sign CRLs.

keyEncipherment :: Maybe Bool

Key can be used to encipher data.

encipherOnly :: Maybe Bool

Key can be used only to encipher data.

keyAgreement :: Maybe Bool

Key can be used in a key-agreement protocol.

digitalSignature :: Maybe Bool

Key can be used for digital signing.

decipherOnly :: Maybe Bool

Key can be used only to decipher data.

nonRepudiation :: Maybe Bool

Key can be used for non-repudiation.

newKeyUsage :: KeyUsage

Create a value of KeyUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataEncipherment:KeyUsage', keyUsage_dataEncipherment - Key can be used to decipher data.

$sel:keyCertSign:KeyUsage', keyUsage_keyCertSign - Key can be used to sign certificates.

$sel:cRLSign:KeyUsage', keyUsage_cRLSign - Key can be used to sign CRLs.

$sel:keyEncipherment:KeyUsage', keyUsage_keyEncipherment - Key can be used to encipher data.

$sel:encipherOnly:KeyUsage', keyUsage_encipherOnly - Key can be used only to encipher data.

$sel:keyAgreement:KeyUsage', keyUsage_keyAgreement - Key can be used in a key-agreement protocol.

$sel:digitalSignature:KeyUsage', keyUsage_digitalSignature - Key can be used for digital signing.

$sel:decipherOnly:KeyUsage', keyUsage_decipherOnly - Key can be used only to decipher data.

$sel:nonRepudiation:KeyUsage', keyUsage_nonRepudiation - Key can be used for non-repudiation.

keyUsage_dataEncipherment :: Lens' KeyUsage (Maybe Bool)

Key can be used to decipher data.

keyUsage_keyCertSign :: Lens' KeyUsage (Maybe Bool)

Key can be used to sign certificates.

keyUsage_cRLSign :: Lens' KeyUsage (Maybe Bool)

Key can be used to sign CRLs.

keyUsage_keyEncipherment :: Lens' KeyUsage (Maybe Bool)

Key can be used to encipher data.

keyUsage_encipherOnly :: Lens' KeyUsage (Maybe Bool)

Key can be used only to encipher data.

keyUsage_keyAgreement :: Lens' KeyUsage (Maybe Bool)

Key can be used in a key-agreement protocol.

keyUsage_digitalSignature :: Lens' KeyUsage (Maybe Bool)

Key can be used for digital signing.

keyUsage_decipherOnly :: Lens' KeyUsage (Maybe Bool)

Key can be used only to decipher data.

keyUsage_nonRepudiation :: Lens' KeyUsage (Maybe Bool)

Key can be used for non-repudiation.