Network.AWS.CertificateManagerPCA.Types.Permission

Description

Documentation

data Permission

Permissions designate which private CA actions can be performed by an AWS service or entity. In order for ACM to automatically renew private certificates, you must give the ACM service principal all available permissions (IssueCertificate, GetCertificate, and ListPermissions). Permissions can be assigned with the CreatePermission action, removed with the DeletePermission action, and listed with the ListPermissions action.

See: newPermission smart constructor.

Constructors

Permission'

Fields

certificateAuthorityArn :: Maybe Text

The Amazon Resource Number (ARN) of the private CA from which the permission was issued.

createdAt :: Maybe POSIX

The time at which the permission was created.

actions :: Maybe (NonEmpty ActionType)

The private CA actions that can be performed by the designated AWS service.

principal :: Maybe Text

The AWS service or entity that holds the permission. At this time, the only valid principal is acm.amazonaws.com.

sourceAccount :: Maybe Text

The ID of the account that assigned the permission.

policy :: Maybe Text

The name of the policy that is associated with the permission.

newPermission :: Permission

Create a value of Permission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateAuthorityArn:Permission', permission_certificateAuthorityArn - The Amazon Resource Number (ARN) of the private CA from which the permission was issued.

$sel:createdAt:Permission', permission_createdAt - The time at which the permission was created.

$sel:actions:Permission', permission_actions - The private CA actions that can be performed by the designated AWS service.

$sel:principal:Permission', permission_principal - The AWS service or entity that holds the permission. At this time, the only valid principal is acm.amazonaws.com.

$sel:sourceAccount:Permission', permission_sourceAccount - The ID of the account that assigned the permission.

$sel:policy:Permission', permission_policy - The name of the policy that is associated with the permission.

permission_certificateAuthorityArn :: Lens' Permission (Maybe Text)

The Amazon Resource Number (ARN) of the private CA from which the permission was issued.

permission_createdAt :: Lens' Permission (Maybe UTCTime)

The time at which the permission was created.

permission_actions :: Lens' Permission (Maybe (NonEmpty ActionType))

The private CA actions that can be performed by the designated AWS service.

permission_principal :: Lens' Permission (Maybe Text)

The AWS service or entity that holds the permission. At this time, the only valid principal is acm.amazonaws.com.

permission_sourceAccount :: Lens' Permission (Maybe Text)

The ID of the account that assigned the permission.

permission_policy :: Lens' Permission (Maybe Text)

The name of the policy that is associated with the permission.