Network.AWS.Chime.BatchCreateAttendee

Description

Creates up to 100 new attendees for an active Amazon Chime SDK meeting. For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide.

Creating a Request

data BatchCreateAttendee

See: newBatchCreateAttendee smart constructor.

Constructors

BatchCreateAttendee'

Fields

meetingId :: Text

The Amazon Chime SDK meeting ID.

attendees :: [CreateAttendeeRequestItem]

The request containing the attendees to create.

newBatchCreateAttendee

Arguments

:: Text
$sel:meetingId:BatchCreateAttendee'

-> BatchCreateAttendee

Create a value of BatchCreateAttendee with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meetingId:BatchCreateAttendee', batchCreateAttendee_meetingId - The Amazon Chime SDK meeting ID.

$sel:attendees:BatchCreateAttendee', batchCreateAttendee_attendees - The request containing the attendees to create.

Request Lenses

batchCreateAttendee_meetingId :: Lens' BatchCreateAttendee Text

The Amazon Chime SDK meeting ID.

batchCreateAttendee_attendees :: Lens' BatchCreateAttendee [CreateAttendeeRequestItem]

The request containing the attendees to create.

Destructuring the Response

data BatchCreateAttendeeResponse

See: newBatchCreateAttendeeResponse smart constructor.

Constructors

BatchCreateAttendeeResponse'

Fields

attendees :: Maybe [Attendee]

The attendee information, including attendees IDs and join tokens.

errors :: Maybe [CreateAttendeeError]

If the action fails for one or more of the attendees in the request, a list of the attendees is returned, along with error codes and error messages.

httpStatus :: Int

The response's http status code.

newBatchCreateAttendeeResponse

Arguments

:: Int
$sel:httpStatus:BatchCreateAttendeeResponse'

-> BatchCreateAttendeeResponse

Create a value of BatchCreateAttendeeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attendees:BatchCreateAttendee', batchCreateAttendeeResponse_attendees - The attendee information, including attendees IDs and join tokens.

$sel:errors:BatchCreateAttendeeResponse', batchCreateAttendeeResponse_errors - If the action fails for one or more of the attendees in the request, a list of the attendees is returned, along with error codes and error messages.

$sel:httpStatus:BatchCreateAttendeeResponse', batchCreateAttendeeResponse_httpStatus - The response's http status code.

Response Lenses

batchCreateAttendeeResponse_attendees :: Lens' BatchCreateAttendeeResponse (Maybe [Attendee])

The attendee information, including attendees IDs and join tokens.

batchCreateAttendeeResponse_errors :: Lens' BatchCreateAttendeeResponse (Maybe [CreateAttendeeError])

If the action fails for one or more of the attendees in the request, a list of the attendees is returned, along with error codes and error messages.

batchCreateAttendeeResponse_httpStatus :: Lens' BatchCreateAttendeeResponse Int

The response's http status code.