Network.AWS.Chime.BatchCreateChannelMembership

Description

Adds a specified number of users to a channel.

Creating a Request

data BatchCreateChannelMembership

See: newBatchCreateChannelMembership smart constructor.

Constructors

BatchCreateChannelMembership'

Fields

chimeBearer :: Maybe Text

The AppInstanceUserArn of the user that makes the API call.

type' :: Maybe ChannelMembershipType

The membership type of a user, DEFAULT or HIDDEN. Default members are always returned as part of ListChannelMemberships. Hidden members are only returned if the type filter in ListChannelMemberships equals HIDDEN. Otherwise hidden members are not returned. This is only supported by moderators.

channelArn :: Text

The ARN of the channel to which you're adding users.

memberArns :: NonEmpty Text

The ARNs of the members you want to add to the channel.

newBatchCreateChannelMembership

Arguments

:: Text
$sel:channelArn:BatchCreateChannelMembership'

-> NonEmpty Text
$sel:memberArns:BatchCreateChannelMembership'

-> BatchCreateChannelMembership

Create a value of BatchCreateChannelMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:chimeBearer:BatchCreateChannelMembership', batchCreateChannelMembership_chimeBearer - The AppInstanceUserArn of the user that makes the API call.

$sel:type':BatchCreateChannelMembership', batchCreateChannelMembership_type - The membership type of a user, DEFAULT or HIDDEN. Default members are always returned as part of ListChannelMemberships. Hidden members are only returned if the type filter in ListChannelMemberships equals HIDDEN. Otherwise hidden members are not returned. This is only supported by moderators.

$sel:channelArn:BatchCreateChannelMembership', batchCreateChannelMembership_channelArn - The ARN of the channel to which you're adding users.

$sel:memberArns:BatchCreateChannelMembership', batchCreateChannelMembership_memberArns - The ARNs of the members you want to add to the channel.

Request Lenses

batchCreateChannelMembership_chimeBearer :: Lens' BatchCreateChannelMembership (Maybe Text)

The AppInstanceUserArn of the user that makes the API call.

batchCreateChannelMembership_type :: Lens' BatchCreateChannelMembership (Maybe ChannelMembershipType)

The membership type of a user, DEFAULT or HIDDEN. Default members are always returned as part of ListChannelMemberships. Hidden members are only returned if the type filter in ListChannelMemberships equals HIDDEN. Otherwise hidden members are not returned. This is only supported by moderators.

batchCreateChannelMembership_channelArn :: Lens' BatchCreateChannelMembership Text

The ARN of the channel to which you're adding users.

batchCreateChannelMembership_memberArns :: Lens' BatchCreateChannelMembership (NonEmpty Text)

The ARNs of the members you want to add to the channel.

Destructuring the Response

data BatchCreateChannelMembershipResponse

See: newBatchCreateChannelMembershipResponse smart constructor.

Constructors

BatchCreateChannelMembershipResponse'

Fields

errors :: Maybe [BatchCreateChannelMembershipError]

If the action fails for one or more of the memberships in the request, a list of the memberships is returned, along with error codes and error messages.

batchChannelMemberships :: Maybe BatchChannelMemberships

The list of channel memberships in the response.

httpStatus :: Int

The response's http status code.

newBatchCreateChannelMembershipResponse

Arguments

:: Int
$sel:httpStatus:BatchCreateChannelMembershipResponse'

-> BatchCreateChannelMembershipResponse

Create a value of BatchCreateChannelMembershipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errors:BatchCreateChannelMembershipResponse', batchCreateChannelMembershipResponse_errors - If the action fails for one or more of the memberships in the request, a list of the memberships is returned, along with error codes and error messages.

$sel:batchChannelMemberships:BatchCreateChannelMembershipResponse', batchCreateChannelMembershipResponse_batchChannelMemberships - The list of channel memberships in the response.

$sel:httpStatus:BatchCreateChannelMembershipResponse', batchCreateChannelMembershipResponse_httpStatus - The response's http status code.

Response Lenses

batchCreateChannelMembershipResponse_errors :: Lens' BatchCreateChannelMembershipResponse (Maybe [BatchCreateChannelMembershipError])

If the action fails for one or more of the memberships in the request, a list of the memberships is returned, along with error codes and error messages.

batchCreateChannelMembershipResponse_batchChannelMemberships :: Lens' BatchCreateChannelMembershipResponse (Maybe BatchChannelMemberships)

The list of channel memberships in the response.

batchCreateChannelMembershipResponse_httpStatus :: Lens' BatchCreateChannelMembershipResponse Int

The response's http status code.