Network.AWS.Chime.BatchSuspendUser

Description

Suspends up to 50 users from a Team or EnterpriseLWA Amazon Chime account. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

Users suspended from a Team account are disassociated from the account,but they can continue to use Amazon Chime as free users. To remove the suspension from suspended Team account users, invite them to the Team account again. You can use the InviteUsers action to do so.

Users suspended from an EnterpriseLWA account are immediately signed out of Amazon Chime and can no longer sign in. To remove the suspension from suspended EnterpriseLWA account users, use the BatchUnsuspendUser action.

To sign out users without suspending them, use the LogoutUser action.

Creating a Request

data BatchSuspendUser

See: newBatchSuspendUser smart constructor.

Constructors

BatchSuspendUser'

Fields

accountId :: Text

The Amazon Chime account ID.

userIdList :: [Text]

The request containing the user IDs to suspend.

newBatchSuspendUser

Arguments

:: Text
$sel:accountId:BatchSuspendUser'

-> BatchSuspendUser

Create a value of BatchSuspendUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:BatchSuspendUser', batchSuspendUser_accountId - The Amazon Chime account ID.

$sel:userIdList:BatchSuspendUser', batchSuspendUser_userIdList - The request containing the user IDs to suspend.

Request Lenses

batchSuspendUser_accountId :: Lens' BatchSuspendUser Text

The Amazon Chime account ID.

batchSuspendUser_userIdList :: Lens' BatchSuspendUser [Text]

The request containing the user IDs to suspend.

Destructuring the Response

data BatchSuspendUserResponse

See: newBatchSuspendUserResponse smart constructor.

Constructors

BatchSuspendUserResponse'

Fields

userErrors :: Maybe [UserError]

If the BatchSuspendUser action fails for one or more of the user IDs in the request, a list of the user IDs is returned, along with error codes and error messages.

httpStatus :: Int

The response's http status code.

newBatchSuspendUserResponse

Arguments

:: Int
$sel:httpStatus:BatchSuspendUserResponse'

-> BatchSuspendUserResponse

Create a value of BatchSuspendUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userErrors:BatchSuspendUserResponse', batchSuspendUserResponse_userErrors - If the BatchSuspendUser action fails for one or more of the user IDs in the request, a list of the user IDs is returned, along with error codes and error messages.

$sel:httpStatus:BatchSuspendUserResponse', batchSuspendUserResponse_httpStatus - The response's http status code.

Response Lenses

batchSuspendUserResponse_userErrors :: Lens' BatchSuspendUserResponse (Maybe [UserError])

If the BatchSuspendUser action fails for one or more of the user IDs in the request, a list of the user IDs is returned, along with error codes and error messages.

batchSuspendUserResponse_httpStatus :: Lens' BatchSuspendUserResponse Int

The response's http status code.