Network.AWS.Chime.CreateChannel

Description

Creates a channel to which you can add users and send messages.

Restriction: You can't change a channel's privacy.

The x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Creating a Request

data CreateChannel

See: newCreateChannel smart constructor.

Constructors

CreateChannel'

Fields

mode :: Maybe ChannelMode

The channel mode: UNRESTRICTED or RESTRICTED. Administrators, moderators, and channel members can add themselves and other members to unrestricted channels. Only administrators and moderators can add members to restricted channels.

privacy :: Maybe ChannelPrivacy

The channel's privacy level: PUBLIC or PRIVATE. Private channels aren't discoverable by users outside the channel. Public channels are discoverable by anyone in the AppInstance.

chimeBearer :: Maybe Text

The AppInstanceUserArn of the user that makes the API call.

metadata :: Maybe (Sensitive Text)

The metadata of the creation request. Limited to 1KB and UTF-8.

tags :: Maybe (NonEmpty Tag)

The tags for the creation request.

appInstanceArn :: Text

The ARN of the channel request.

name :: Sensitive Text

The name of the channel.

clientRequestToken :: Sensitive Text

The client token for the request. An Idempotency token.

newCreateChannel

Arguments

:: Text
$sel:appInstanceArn:CreateChannel'

-> Text
$sel:name:CreateChannel'

-> Text
$sel:clientRequestToken:CreateChannel'

-> CreateChannel

Create a value of CreateChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mode:CreateChannel', createChannel_mode - The channel mode: UNRESTRICTED or RESTRICTED. Administrators, moderators, and channel members can add themselves and other members to unrestricted channels. Only administrators and moderators can add members to restricted channels.

$sel:privacy:CreateChannel', createChannel_privacy - The channel's privacy level: PUBLIC or PRIVATE. Private channels aren't discoverable by users outside the channel. Public channels are discoverable by anyone in the AppInstance.

$sel:chimeBearer:CreateChannel', createChannel_chimeBearer - The AppInstanceUserArn of the user that makes the API call.

$sel:metadata:CreateChannel', createChannel_metadata - The metadata of the creation request. Limited to 1KB and UTF-8.

$sel:tags:CreateChannel', createChannel_tags - The tags for the creation request.

$sel:appInstanceArn:CreateChannel', createChannel_appInstanceArn - The ARN of the channel request.

$sel:name:CreateChannel', createChannel_name - The name of the channel.

$sel:clientRequestToken:CreateChannel', createChannel_clientRequestToken - The client token for the request. An Idempotency token.

Request Lenses

createChannel_mode :: Lens' CreateChannel (Maybe ChannelMode)

The channel mode: UNRESTRICTED or RESTRICTED. Administrators, moderators, and channel members can add themselves and other members to unrestricted channels. Only administrators and moderators can add members to restricted channels.

createChannel_privacy :: Lens' CreateChannel (Maybe ChannelPrivacy)

The channel's privacy level: PUBLIC or PRIVATE. Private channels aren't discoverable by users outside the channel. Public channels are discoverable by anyone in the AppInstance.

createChannel_chimeBearer :: Lens' CreateChannel (Maybe Text)

The AppInstanceUserArn of the user that makes the API call.

createChannel_metadata :: Lens' CreateChannel (Maybe Text)

The metadata of the creation request. Limited to 1KB and UTF-8.

createChannel_tags :: Lens' CreateChannel (Maybe (NonEmpty Tag))

The tags for the creation request.

createChannel_appInstanceArn :: Lens' CreateChannel Text

The ARN of the channel request.

createChannel_name :: Lens' CreateChannel Text

The name of the channel.

createChannel_clientRequestToken :: Lens' CreateChannel Text

The client token for the request. An Idempotency token.

Destructuring the Response

data CreateChannelResponse

See: newCreateChannelResponse smart constructor.

Constructors

CreateChannelResponse'

Fields

channelArn :: Maybe Text

The ARN of the channel.

httpStatus :: Int

The response's http status code.

newCreateChannelResponse

Arguments

:: Int
$sel:httpStatus:CreateChannelResponse'

-> CreateChannelResponse

Create a value of CreateChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:CreateChannelResponse', createChannelResponse_channelArn - The ARN of the channel.

$sel:httpStatus:CreateChannelResponse', createChannelResponse_httpStatus - The response's http status code.

Response Lenses

createChannelResponse_channelArn :: Lens' CreateChannelResponse (Maybe Text)

The ARN of the channel.

createChannelResponse_httpStatus :: Lens' CreateChannelResponse Int

The response's http status code.