Network.AWS.Chime.CreateMeetingDialOut

Description

Uses the join token and call metadata in a meeting request (From number, To number, and so forth) to initiate an outbound call to a public switched telephone network (PSTN) and join them into a Chime meeting. Also ensures that the From number belongs to the customer.

To play welcome audio or implement an interactive voice response (IVR), use the CreateSipMediaApplicationCall action with the corresponding SIP media application ID.

Creating a Request

data CreateMeetingDialOut

See: newCreateMeetingDialOut smart constructor.

Constructors

CreateMeetingDialOut'

Fields

meetingId :: Text

The Amazon Chime SDK meeting ID.

fromPhoneNumber :: Sensitive Text

Phone number used as the caller ID when the remote party receives a call.

toPhoneNumber :: Sensitive Text

Phone number called when inviting someone to a meeting.

joinToken :: Sensitive Text

Token used by the Amazon Chime SDK attendee. Call the CreateAttendee action to get a join token.

newCreateMeetingDialOut

Arguments

:: Text
$sel:meetingId:CreateMeetingDialOut'

-> Text
$sel:fromPhoneNumber:CreateMeetingDialOut'

-> Text
$sel:toPhoneNumber:CreateMeetingDialOut'

-> Text
$sel:joinToken:CreateMeetingDialOut'

-> CreateMeetingDialOut

Create a value of CreateMeetingDialOut with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meetingId:CreateMeetingDialOut', createMeetingDialOut_meetingId - The Amazon Chime SDK meeting ID.

$sel:fromPhoneNumber:CreateMeetingDialOut', createMeetingDialOut_fromPhoneNumber - Phone number used as the caller ID when the remote party receives a call.

$sel:toPhoneNumber:CreateMeetingDialOut', createMeetingDialOut_toPhoneNumber - Phone number called when inviting someone to a meeting.

$sel:joinToken:CreateMeetingDialOut', createMeetingDialOut_joinToken - Token used by the Amazon Chime SDK attendee. Call the CreateAttendee action to get a join token.

Request Lenses

createMeetingDialOut_meetingId :: Lens' CreateMeetingDialOut Text

The Amazon Chime SDK meeting ID.

createMeetingDialOut_fromPhoneNumber :: Lens' CreateMeetingDialOut Text

Phone number used as the caller ID when the remote party receives a call.

createMeetingDialOut_toPhoneNumber :: Lens' CreateMeetingDialOut Text

Phone number called when inviting someone to a meeting.

createMeetingDialOut_joinToken :: Lens' CreateMeetingDialOut Text

Token used by the Amazon Chime SDK attendee. Call the CreateAttendee action to get a join token.

Destructuring the Response

data CreateMeetingDialOutResponse

See: newCreateMeetingDialOutResponse smart constructor.

Constructors

CreateMeetingDialOutResponse'

Fields

transactionId :: Maybe Text

Unique ID that tracks API calls.

httpStatus :: Int

The response's http status code.

newCreateMeetingDialOutResponse

Arguments

:: Int
$sel:httpStatus:CreateMeetingDialOutResponse'

-> CreateMeetingDialOutResponse

Create a value of CreateMeetingDialOutResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transactionId:CreateMeetingDialOutResponse', createMeetingDialOutResponse_transactionId - Unique ID that tracks API calls.

$sel:httpStatus:CreateMeetingDialOutResponse', createMeetingDialOutResponse_httpStatus - The response's http status code.

Response Lenses

createMeetingDialOutResponse_transactionId :: Lens' CreateMeetingDialOutResponse (Maybe Text)

Unique ID that tracks API calls.

createMeetingDialOutResponse_httpStatus :: Lens' CreateMeetingDialOutResponse Int

The response's http status code.