Network.AWS.Chime.CreateMeetingWithAttendees

Description

Creates a new Amazon Chime SDK meeting in the specified media Region, with attendees. For more information about specifying media Regions, see Amazon Chime SDK Media Regions in the Amazon Chime Developer Guide . For more information about the Amazon Chime SDK, see Using the Amazon Chime SDK in the Amazon Chime Developer Guide .

Creating a Request

data CreateMeetingWithAttendees

See: newCreateMeetingWithAttendees smart constructor.

Constructors

CreateMeetingWithAttendees'

Fields

mediaRegion :: Maybe Text

The Region in which to create the meeting. Default: us-east-1 .

Available values: af-south-1 , ap-northeast-1 , ap-northeast-2 , ap-south-1 , ap-southeast-1 , ap-southeast-2 , ca-central-1 , eu-central-1 , eu-north-1 , eu-south-1 , eu-west-1 , eu-west-2 , eu-west-3 , sa-east-1 , us-east-1 , us-east-2 , us-west-1 , us-west-2 .

meetingHostId :: Maybe (Sensitive Text)

Reserved.

attendees :: Maybe (NonEmpty CreateAttendeeRequestItem)

The request containing the attendees to create.

notificationsConfiguration :: Maybe MeetingNotificationConfiguration

externalMeetingId :: Maybe (Sensitive Text)

The external meeting ID.

tags :: Maybe (NonEmpty Tag)

The tag key-value pairs.

clientRequestToken :: Sensitive Text

The unique identifier for the client request. Use a different token for different meetings.

newCreateMeetingWithAttendees

Arguments

:: Text
$sel:clientRequestToken:CreateMeetingWithAttendees'

-> CreateMeetingWithAttendees

Create a value of CreateMeetingWithAttendees with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mediaRegion:CreateMeetingWithAttendees', createMeetingWithAttendees_mediaRegion - The Region in which to create the meeting. Default: us-east-1 .

Available values: af-south-1 , ap-northeast-1 , ap-northeast-2 , ap-south-1 , ap-southeast-1 , ap-southeast-2 , ca-central-1 , eu-central-1 , eu-north-1 , eu-south-1 , eu-west-1 , eu-west-2 , eu-west-3 , sa-east-1 , us-east-1 , us-east-2 , us-west-1 , us-west-2 .

$sel:meetingHostId:CreateMeetingWithAttendees', createMeetingWithAttendees_meetingHostId - Reserved.

$sel:attendees:CreateMeetingWithAttendees', createMeetingWithAttendees_attendees - The request containing the attendees to create.

$sel:notificationsConfiguration:CreateMeetingWithAttendees', createMeetingWithAttendees_notificationsConfiguration - Undocumented member.

$sel:externalMeetingId:CreateMeetingWithAttendees', createMeetingWithAttendees_externalMeetingId - The external meeting ID.

$sel:tags:CreateMeetingWithAttendees', createMeetingWithAttendees_tags - The tag key-value pairs.

$sel:clientRequestToken:CreateMeetingWithAttendees', createMeetingWithAttendees_clientRequestToken - The unique identifier for the client request. Use a different token for different meetings.

Request Lenses

createMeetingWithAttendees_mediaRegion :: Lens' CreateMeetingWithAttendees (Maybe Text)

The Region in which to create the meeting. Default: us-east-1 .

Available values: af-south-1 , ap-northeast-1 , ap-northeast-2 , ap-south-1 , ap-southeast-1 , ap-southeast-2 , ca-central-1 , eu-central-1 , eu-north-1 , eu-south-1 , eu-west-1 , eu-west-2 , eu-west-3 , sa-east-1 , us-east-1 , us-east-2 , us-west-1 , us-west-2 .

createMeetingWithAttendees_meetingHostId :: Lens' CreateMeetingWithAttendees (Maybe Text)

Reserved.

createMeetingWithAttendees_attendees :: Lens' CreateMeetingWithAttendees (Maybe (NonEmpty CreateAttendeeRequestItem))

The request containing the attendees to create.

createMeetingWithAttendees_notificationsConfiguration :: Lens' CreateMeetingWithAttendees (Maybe MeetingNotificationConfiguration)

Undocumented member.

createMeetingWithAttendees_externalMeetingId :: Lens' CreateMeetingWithAttendees (Maybe Text)

The external meeting ID.

createMeetingWithAttendees_tags :: Lens' CreateMeetingWithAttendees (Maybe (NonEmpty Tag))

The tag key-value pairs.

createMeetingWithAttendees_clientRequestToken :: Lens' CreateMeetingWithAttendees Text

The unique identifier for the client request. Use a different token for different meetings.

Destructuring the Response

data CreateMeetingWithAttendeesResponse

See: newCreateMeetingWithAttendeesResponse smart constructor.

Constructors

CreateMeetingWithAttendeesResponse'

Fields

attendees :: Maybe [Attendee]

The attendee information, including attendees IDs and join tokens.

meeting :: Maybe Meeting

errors :: Maybe [CreateAttendeeError]

If the action fails for one or more of the attendees in the request, a list of the attendees is returned, along with error codes and error messages.

httpStatus :: Int

The response's http status code.

newCreateMeetingWithAttendeesResponse

Arguments

:: Int
$sel:httpStatus:CreateMeetingWithAttendeesResponse'

-> CreateMeetingWithAttendeesResponse

Create a value of CreateMeetingWithAttendeesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attendees:CreateMeetingWithAttendees', createMeetingWithAttendeesResponse_attendees - The attendee information, including attendees IDs and join tokens.

$sel:meeting:CreateMeetingWithAttendeesResponse', createMeetingWithAttendeesResponse_meeting - Undocumented member.

$sel:errors:CreateMeetingWithAttendeesResponse', createMeetingWithAttendeesResponse_errors - If the action fails for one or more of the attendees in the request, a list of the attendees is returned, along with error codes and error messages.

$sel:httpStatus:CreateMeetingWithAttendeesResponse', createMeetingWithAttendeesResponse_httpStatus - The response's http status code.

Response Lenses

createMeetingWithAttendeesResponse_attendees :: Lens' CreateMeetingWithAttendeesResponse (Maybe [Attendee])

The attendee information, including attendees IDs and join tokens.

createMeetingWithAttendeesResponse_meeting :: Lens' CreateMeetingWithAttendeesResponse (Maybe Meeting)

Undocumented member.

createMeetingWithAttendeesResponse_errors :: Lens' CreateMeetingWithAttendeesResponse (Maybe [CreateAttendeeError])

If the action fails for one or more of the attendees in the request, a list of the attendees is returned, along with error codes and error messages.

createMeetingWithAttendeesResponse_httpStatus :: Lens' CreateMeetingWithAttendeesResponse Int

The response's http status code.