Network.AWS.Chime.CreateProxySession

Description

Creates a proxy session on the specified Amazon Chime Voice Connector for the specified participant phone numbers.

Creating a Request

data CreateProxySession

See: newCreateProxySession smart constructor.

Constructors

CreateProxySession'

Fields

numberSelectionBehavior :: Maybe NumberSelectionBehavior

The preference for proxy phone number reuse, or stickiness, between the same participants across sessions.

geoMatchParams :: Maybe GeoMatchParams

The country and area code for the proxy phone number.

expiryMinutes :: Maybe Natural

The number of minutes allowed for the proxy session.

name :: Maybe (Sensitive Text)

The name of the proxy session.

geoMatchLevel :: Maybe GeoMatchLevel

The preference for matching the country or area code of the proxy phone number with that of the first participant.

participantPhoneNumbers :: NonEmpty (Sensitive Text)

The participant phone numbers.

capabilities :: [Capability]

The proxy session capabilities.

voiceConnectorId :: Text

The Amazon Chime voice connector ID.

newCreateProxySession

Arguments

:: NonEmpty Text
$sel:participantPhoneNumbers:CreateProxySession'

-> Text
$sel:voiceConnectorId:CreateProxySession'

-> CreateProxySession

Create a value of CreateProxySession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numberSelectionBehavior:CreateProxySession', createProxySession_numberSelectionBehavior - The preference for proxy phone number reuse, or stickiness, between the same participants across sessions.

$sel:geoMatchParams:CreateProxySession', createProxySession_geoMatchParams - The country and area code for the proxy phone number.

$sel:expiryMinutes:CreateProxySession', createProxySession_expiryMinutes - The number of minutes allowed for the proxy session.

$sel:name:CreateProxySession', createProxySession_name - The name of the proxy session.

$sel:geoMatchLevel:CreateProxySession', createProxySession_geoMatchLevel - The preference for matching the country or area code of the proxy phone number with that of the first participant.

$sel:participantPhoneNumbers:CreateProxySession', createProxySession_participantPhoneNumbers - The participant phone numbers.

$sel:capabilities:CreateProxySession', createProxySession_capabilities - The proxy session capabilities.

$sel:voiceConnectorId:CreateProxySession', createProxySession_voiceConnectorId - The Amazon Chime voice connector ID.

Request Lenses

createProxySession_numberSelectionBehavior :: Lens' CreateProxySession (Maybe NumberSelectionBehavior)

The preference for proxy phone number reuse, or stickiness, between the same participants across sessions.

createProxySession_geoMatchParams :: Lens' CreateProxySession (Maybe GeoMatchParams)

The country and area code for the proxy phone number.

createProxySession_expiryMinutes :: Lens' CreateProxySession (Maybe Natural)

The number of minutes allowed for the proxy session.

createProxySession_name :: Lens' CreateProxySession (Maybe Text)

The name of the proxy session.

createProxySession_geoMatchLevel :: Lens' CreateProxySession (Maybe GeoMatchLevel)

The preference for matching the country or area code of the proxy phone number with that of the first participant.

createProxySession_participantPhoneNumbers :: Lens' CreateProxySession (NonEmpty Text)

The participant phone numbers.

createProxySession_capabilities :: Lens' CreateProxySession [Capability]

The proxy session capabilities.

createProxySession_voiceConnectorId :: Lens' CreateProxySession Text

The Amazon Chime voice connector ID.

Destructuring the Response

data CreateProxySessionResponse

See: newCreateProxySessionResponse smart constructor.

Constructors

CreateProxySessionResponse'

Fields

proxySession :: Maybe ProxySession

The proxy session details.

httpStatus :: Int

The response's http status code.

newCreateProxySessionResponse

Arguments

:: Int
$sel:httpStatus:CreateProxySessionResponse'

-> CreateProxySessionResponse

Create a value of CreateProxySessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:proxySession:CreateProxySessionResponse', createProxySessionResponse_proxySession - The proxy session details.

$sel:httpStatus:CreateProxySessionResponse', createProxySessionResponse_httpStatus - The response's http status code.

Response Lenses

createProxySessionResponse_proxySession :: Lens' CreateProxySessionResponse (Maybe ProxySession)

The proxy session details.

createProxySessionResponse_httpStatus :: Lens' CreateProxySessionResponse Int

The response's http status code.