Network.AWS.Chime.CreateSipRule

Description

Creates a SIP rule which can be used to run a SIP media application as a target for a specific trigger type.

Creating a Request

data CreateSipRule

See: newCreateSipRule smart constructor.

Constructors

CreateSipRule'

Fields

disabled :: Maybe Bool

Enables or disables a rule. You must disable rules before you can delete them.

name :: Text

The name of the SIP rule.

triggerType :: SipRuleTriggerType

The type of trigger assigned to the SIP rule in TriggerValue, currently RequestUriHostname or ToPhoneNumber.

triggerValue :: Text

If TriggerType is RequestUriHostname, the value can be the outbound host name of an Amazon Chime Voice Connector. If TriggerType is ToPhoneNumber, the value can be a customer-owned phone number in the E164 format. The SipMediaApplication specified in the SipRule is triggered if the request URI in an incoming SIP request matches the RequestUriHostname, or if the To header in the incoming SIP request matches the ToPhoneNumber value.

targetApplications :: NonEmpty SipRuleTargetApplication

List of SIP media applications with priority and AWS Region. Only one SIP application per AWS Region can be used.

newCreateSipRule

Arguments

:: Text
$sel:name:CreateSipRule'

-> SipRuleTriggerType
$sel:triggerType:CreateSipRule'

-> Text
$sel:triggerValue:CreateSipRule'

-> NonEmpty SipRuleTargetApplication
$sel:targetApplications:CreateSipRule'

-> CreateSipRule

Create a value of CreateSipRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disabled:CreateSipRule', createSipRule_disabled - Enables or disables a rule. You must disable rules before you can delete them.

$sel:name:CreateSipRule', createSipRule_name - The name of the SIP rule.

$sel:triggerType:CreateSipRule', createSipRule_triggerType - The type of trigger assigned to the SIP rule in TriggerValue, currently RequestUriHostname or ToPhoneNumber.

$sel:triggerValue:CreateSipRule', createSipRule_triggerValue - If TriggerType is RequestUriHostname, the value can be the outbound host name of an Amazon Chime Voice Connector. If TriggerType is ToPhoneNumber, the value can be a customer-owned phone number in the E164 format. The SipMediaApplication specified in the SipRule is triggered if the request URI in an incoming SIP request matches the RequestUriHostname, or if the To header in the incoming SIP request matches the ToPhoneNumber value.

$sel:targetApplications:CreateSipRule', createSipRule_targetApplications - List of SIP media applications with priority and AWS Region. Only one SIP application per AWS Region can be used.

Request Lenses

createSipRule_disabled :: Lens' CreateSipRule (Maybe Bool)

Enables or disables a rule. You must disable rules before you can delete them.

createSipRule_name :: Lens' CreateSipRule Text

The name of the SIP rule.

createSipRule_triggerType :: Lens' CreateSipRule SipRuleTriggerType

The type of trigger assigned to the SIP rule in TriggerValue, currently RequestUriHostname or ToPhoneNumber.

createSipRule_triggerValue :: Lens' CreateSipRule Text

If TriggerType is RequestUriHostname, the value can be the outbound host name of an Amazon Chime Voice Connector. If TriggerType is ToPhoneNumber, the value can be a customer-owned phone number in the E164 format. The SipMediaApplication specified in the SipRule is triggered if the request URI in an incoming SIP request matches the RequestUriHostname, or if the To header in the incoming SIP request matches the ToPhoneNumber value.

createSipRule_targetApplications :: Lens' CreateSipRule (NonEmpty SipRuleTargetApplication)

List of SIP media applications with priority and AWS Region. Only one SIP application per AWS Region can be used.

Destructuring the Response

data CreateSipRuleResponse

See: newCreateSipRuleResponse smart constructor.

Constructors

CreateSipRuleResponse'

Fields

sipRule :: Maybe SipRule

Returns the SIP rule information, including the rule ID, triggers, and target applications.

httpStatus :: Int

The response's http status code.

newCreateSipRuleResponse

Arguments

:: Int
$sel:httpStatus:CreateSipRuleResponse'

-> CreateSipRuleResponse

Create a value of CreateSipRuleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sipRule:CreateSipRuleResponse', createSipRuleResponse_sipRule - Returns the SIP rule information, including the rule ID, triggers, and target applications.

$sel:httpStatus:CreateSipRuleResponse', createSipRuleResponse_httpStatus - The response's http status code.

Response Lenses

createSipRuleResponse_sipRule :: Lens' CreateSipRuleResponse (Maybe SipRule)

Returns the SIP rule information, including the rule ID, triggers, and target applications.

createSipRuleResponse_httpStatus :: Lens' CreateSipRuleResponse Int

The response's http status code.