Network.AWS.Chime.CreateUser

Description

Creates a user under the specified Amazon Chime account.

Creating a Request

data CreateUser

See: newCreateUser smart constructor.

Constructors

CreateUser'

Fields

email :: Maybe (Sensitive Text)

The user's email address.

username :: Maybe Text

The user name.

userType :: Maybe UserType

The user type.

accountId :: Text

The Amazon Chime account ID.

newCreateUser

Arguments

:: Text
$sel:accountId:CreateUser'

-> CreateUser

Create a value of CreateUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:CreateUser', createUser_email - The user's email address.

$sel:username:CreateUser', createUser_username - The user name.

$sel:userType:CreateUser', createUser_userType - The user type.

$sel:accountId:CreateUser', createUser_accountId - The Amazon Chime account ID.

Request Lenses

createUser_email :: Lens' CreateUser (Maybe Text)

The user's email address.

createUser_username :: Lens' CreateUser (Maybe Text)

The user name.

createUser_userType :: Lens' CreateUser (Maybe UserType)

The user type.

createUser_accountId :: Lens' CreateUser Text

The Amazon Chime account ID.

Destructuring the Response

data CreateUserResponse

See: newCreateUserResponse smart constructor.

Constructors

CreateUserResponse'

Fields

user :: Maybe User

httpStatus :: Int

The response's http status code.

newCreateUserResponse

Arguments

:: Int
$sel:httpStatus:CreateUserResponse'

-> CreateUserResponse

Create a value of CreateUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:user:CreateUserResponse', createUserResponse_user - Undocumented member.

$sel:httpStatus:CreateUserResponse', createUserResponse_httpStatus - The response's http status code.

Response Lenses

createUserResponse_user :: Lens' CreateUserResponse (Maybe User)

Undocumented member.

createUserResponse_httpStatus :: Lens' CreateUserResponse Int

The response's http status code.