Network.AWS.Chime.CreateVoiceConnectorGroup

Description

Creates an Amazon Chime Voice Connector group under the administrator's AWS account. You can associate Amazon Chime Voice Connectors with the Amazon Chime Voice Connector group by including VoiceConnectorItems in the request.

You can include Amazon Chime Voice Connectors from different AWS Regions in your group. This creates a fault tolerant mechanism for fallback in case of availability events.

Creating a Request

data CreateVoiceConnectorGroup

See: newCreateVoiceConnectorGroup smart constructor.

Constructors

CreateVoiceConnectorGroup'

Fields

voiceConnectorItems :: Maybe [VoiceConnectorItem]

The Amazon Chime Voice Connectors to route inbound calls to.

name :: Text

The name of the Amazon Chime Voice Connector group.

newCreateVoiceConnectorGroup

Arguments

:: Text
$sel:name:CreateVoiceConnectorGroup'

-> CreateVoiceConnectorGroup

Create a value of CreateVoiceConnectorGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:voiceConnectorItems:CreateVoiceConnectorGroup', createVoiceConnectorGroup_voiceConnectorItems - The Amazon Chime Voice Connectors to route inbound calls to.

$sel:name:CreateVoiceConnectorGroup', createVoiceConnectorGroup_name - The name of the Amazon Chime Voice Connector group.

Request Lenses

createVoiceConnectorGroup_voiceConnectorItems :: Lens' CreateVoiceConnectorGroup (Maybe [VoiceConnectorItem])

The Amazon Chime Voice Connectors to route inbound calls to.

createVoiceConnectorGroup_name :: Lens' CreateVoiceConnectorGroup Text

The name of the Amazon Chime Voice Connector group.

Destructuring the Response

data CreateVoiceConnectorGroupResponse

See: newCreateVoiceConnectorGroupResponse smart constructor.

Constructors

CreateVoiceConnectorGroupResponse'

Fields

voiceConnectorGroup :: Maybe VoiceConnectorGroup

The Amazon Chime Voice Connector group details.

httpStatus :: Int

The response's http status code.

newCreateVoiceConnectorGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateVoiceConnectorGroupResponse'

-> CreateVoiceConnectorGroupResponse

Create a value of CreateVoiceConnectorGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:voiceConnectorGroup:CreateVoiceConnectorGroupResponse', createVoiceConnectorGroupResponse_voiceConnectorGroup - The Amazon Chime Voice Connector group details.

$sel:httpStatus:CreateVoiceConnectorGroupResponse', createVoiceConnectorGroupResponse_httpStatus - The response's http status code.

Response Lenses

createVoiceConnectorGroupResponse_voiceConnectorGroup :: Lens' CreateVoiceConnectorGroupResponse (Maybe VoiceConnectorGroup)

The Amazon Chime Voice Connector group details.

createVoiceConnectorGroupResponse_httpStatus :: Lens' CreateVoiceConnectorGroupResponse Int

The response's http status code.