Network.AWS.Chime.DeleteProxySession

Description

Deletes the specified proxy session from the specified Amazon Chime Voice Connector.

Creating a Request

data DeleteProxySession

See: newDeleteProxySession smart constructor.

Constructors

DeleteProxySession'

Fields

voiceConnectorId :: Text

The Amazon Chime voice connector ID.

proxySessionId :: Text

The proxy session ID.

newDeleteProxySession

Arguments

:: Text
$sel:voiceConnectorId:DeleteProxySession'

-> Text
$sel:proxySessionId:DeleteProxySession'

-> DeleteProxySession

Create a value of DeleteProxySession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:voiceConnectorId:DeleteProxySession', deleteProxySession_voiceConnectorId - The Amazon Chime voice connector ID.

$sel:proxySessionId:DeleteProxySession', deleteProxySession_proxySessionId - The proxy session ID.

Request Lenses

deleteProxySession_voiceConnectorId :: Lens' DeleteProxySession Text

The Amazon Chime voice connector ID.

deleteProxySession_proxySessionId :: Lens' DeleteProxySession Text

The proxy session ID.

Destructuring the Response

data DeleteProxySessionResponse

See: newDeleteProxySessionResponse smart constructor.

Constructors

DeleteProxySessionResponse'

newDeleteProxySessionResponse :: DeleteProxySessionResponse

Create a value of DeleteProxySessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.