Network.AWS.Chime.DeleteRoom

Description

Deletes a chat room in an Amazon Chime Enterprise account.

Creating a Request

data DeleteRoom

See: newDeleteRoom smart constructor.

Constructors

DeleteRoom'

Fields

accountId :: Text

The Amazon Chime account ID.

roomId :: Text

The chat room ID.

newDeleteRoom

Arguments

:: Text
$sel:accountId:DeleteRoom'

-> Text
$sel:roomId:DeleteRoom'

-> DeleteRoom

Create a value of DeleteRoom with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:DeleteRoom', deleteRoom_accountId - The Amazon Chime account ID.

$sel:roomId:DeleteRoom', deleteRoom_roomId - The chat room ID.

Request Lenses

deleteRoom_accountId :: Lens' DeleteRoom Text

The Amazon Chime account ID.

deleteRoom_roomId :: Lens' DeleteRoom Text

The chat room ID.

Destructuring the Response

data DeleteRoomResponse

See: newDeleteRoomResponse smart constructor.

Constructors

DeleteRoomResponse'

newDeleteRoomResponse :: DeleteRoomResponse

Create a value of DeleteRoomResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.