Network.AWS.Chime.DescribeAppInstanceAdmin

Description

Returns the full details of an AppInstanceAdmin.

Creating a Request

data DescribeAppInstanceAdmin

See: newDescribeAppInstanceAdmin smart constructor.

Constructors

DescribeAppInstanceAdmin'

Fields

appInstanceAdminArn :: Text

The ARN of the AppInstanceAdmin.

appInstanceArn :: Text

The ARN of the AppInstance.

newDescribeAppInstanceAdmin

Arguments

:: Text
$sel:appInstanceAdminArn:DescribeAppInstanceAdmin'

-> Text
$sel:appInstanceArn:DescribeAppInstanceAdmin'

-> DescribeAppInstanceAdmin

Create a value of DescribeAppInstanceAdmin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceAdminArn:DescribeAppInstanceAdmin', describeAppInstanceAdmin_appInstanceAdminArn - The ARN of the AppInstanceAdmin.

$sel:appInstanceArn:DescribeAppInstanceAdmin', describeAppInstanceAdmin_appInstanceArn - The ARN of the AppInstance.

Request Lenses

describeAppInstanceAdmin_appInstanceAdminArn :: Lens' DescribeAppInstanceAdmin Text

The ARN of the AppInstanceAdmin.

describeAppInstanceAdmin_appInstanceArn :: Lens' DescribeAppInstanceAdmin Text

The ARN of the AppInstance.

Destructuring the Response

data DescribeAppInstanceAdminResponse

See: newDescribeAppInstanceAdminResponse smart constructor.

Constructors

DescribeAppInstanceAdminResponse'

Fields

appInstanceAdmin :: Maybe AppInstanceAdmin

The ARN and name of the AppInstanceUser, the ARN of the AppInstance, and the created and last-updated timestamps. All timestamps use epoch milliseconds.

httpStatus :: Int

The response's http status code.

newDescribeAppInstanceAdminResponse

Arguments

:: Int
$sel:httpStatus:DescribeAppInstanceAdminResponse'

-> DescribeAppInstanceAdminResponse

Create a value of DescribeAppInstanceAdminResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceAdmin:DescribeAppInstanceAdminResponse', describeAppInstanceAdminResponse_appInstanceAdmin - The ARN and name of the AppInstanceUser, the ARN of the AppInstance, and the created and last-updated timestamps. All timestamps use epoch milliseconds.

$sel:httpStatus:DescribeAppInstanceAdminResponse', describeAppInstanceAdminResponse_httpStatus - The response's http status code.

Response Lenses

describeAppInstanceAdminResponse_appInstanceAdmin :: Lens' DescribeAppInstanceAdminResponse (Maybe AppInstanceAdmin)

The ARN and name of the AppInstanceUser, the ARN of the AppInstance, and the created and last-updated timestamps. All timestamps use epoch milliseconds.

describeAppInstanceAdminResponse_httpStatus :: Lens' DescribeAppInstanceAdminResponse Int

The response's http status code.