Network.AWS.Chime.DisassociatePhoneNumbersFromVoiceConnector

Description

Disassociates the specified phone numbers from the specified Amazon Chime Voice Connector.

Creating a Request

data DisassociatePhoneNumbersFromVoiceConnector

See: newDisassociatePhoneNumbersFromVoiceConnector smart constructor.

Constructors

DisassociatePhoneNumbersFromVoiceConnector'

Fields

voiceConnectorId :: Text

The Amazon Chime Voice Connector ID.

e164PhoneNumbers :: [Sensitive Text]

List of phone numbers, in E.164 format.

newDisassociatePhoneNumbersFromVoiceConnector

Arguments

:: Text
$sel:voiceConnectorId:DisassociatePhoneNumbersFromVoiceConnector'

-> DisassociatePhoneNumbersFromVoiceConnector

Create a value of DisassociatePhoneNumbersFromVoiceConnector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:voiceConnectorId:DisassociatePhoneNumbersFromVoiceConnector', disassociatePhoneNumbersFromVoiceConnector_voiceConnectorId - The Amazon Chime Voice Connector ID.

$sel:e164PhoneNumbers:DisassociatePhoneNumbersFromVoiceConnector', disassociatePhoneNumbersFromVoiceConnector_e164PhoneNumbers - List of phone numbers, in E.164 format.

Request Lenses

disassociatePhoneNumbersFromVoiceConnector_voiceConnectorId :: Lens' DisassociatePhoneNumbersFromVoiceConnector Text

The Amazon Chime Voice Connector ID.

disassociatePhoneNumbersFromVoiceConnector_e164PhoneNumbers :: Lens' DisassociatePhoneNumbersFromVoiceConnector [Text]

List of phone numbers, in E.164 format.

Destructuring the Response

data DisassociatePhoneNumbersFromVoiceConnectorResponse

See: newDisassociatePhoneNumbersFromVoiceConnectorResponse smart constructor.

Constructors

DisassociatePhoneNumbersFromVoiceConnectorResponse'

Fields

phoneNumberErrors :: Maybe [PhoneNumberError]

If the action fails for one or more of the phone numbers in the request, a list of the phone numbers is returned, along with error codes and error messages.

httpStatus :: Int

The response's http status code.

newDisassociatePhoneNumbersFromVoiceConnectorResponse

Arguments

:: Int
$sel:httpStatus:DisassociatePhoneNumbersFromVoiceConnectorResponse'

-> DisassociatePhoneNumbersFromVoiceConnectorResponse

Create a value of DisassociatePhoneNumbersFromVoiceConnectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumberErrors:DisassociatePhoneNumbersFromVoiceConnectorResponse', disassociatePhoneNumbersFromVoiceConnectorResponse_phoneNumberErrors - If the action fails for one or more of the phone numbers in the request, a list of the phone numbers is returned, along with error codes and error messages.

$sel:httpStatus:DisassociatePhoneNumbersFromVoiceConnectorResponse', disassociatePhoneNumbersFromVoiceConnectorResponse_httpStatus - The response's http status code.

Response Lenses

disassociatePhoneNumbersFromVoiceConnectorResponse_phoneNumberErrors :: Lens' DisassociatePhoneNumbersFromVoiceConnectorResponse (Maybe [PhoneNumberError])

If the action fails for one or more of the phone numbers in the request, a list of the phone numbers is returned, along with error codes and error messages.

disassociatePhoneNumbersFromVoiceConnectorResponse_httpStatus :: Lens' DisassociatePhoneNumbersFromVoiceConnectorResponse Int

The response's http status code.