Network.AWS.Chime.GetAppInstanceStreamingConfigurations

Description

Gets the streaming settings for an AppInstance.

Creating a Request

data GetAppInstanceStreamingConfigurations

See: newGetAppInstanceStreamingConfigurations smart constructor.

Constructors

GetAppInstanceStreamingConfigurations'

Fields

appInstanceArn :: Text

The ARN of the AppInstance.

newGetAppInstanceStreamingConfigurations

Arguments

:: Text
$sel:appInstanceArn:GetAppInstanceStreamingConfigurations'

-> GetAppInstanceStreamingConfigurations

Create a value of GetAppInstanceStreamingConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceArn:GetAppInstanceStreamingConfigurations', getAppInstanceStreamingConfigurations_appInstanceArn - The ARN of the AppInstance.

Request Lenses

getAppInstanceStreamingConfigurations_appInstanceArn :: Lens' GetAppInstanceStreamingConfigurations Text

The ARN of the AppInstance.

Destructuring the Response

data GetAppInstanceStreamingConfigurationsResponse

See: newGetAppInstanceStreamingConfigurationsResponse smart constructor.

Constructors

GetAppInstanceStreamingConfigurationsResponse'

Fields

appInstanceStreamingConfigurations :: Maybe (NonEmpty AppInstanceStreamingConfiguration)

The streaming settings.

httpStatus :: Int

The response's http status code.

newGetAppInstanceStreamingConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:GetAppInstanceStreamingConfigurationsResponse'

-> GetAppInstanceStreamingConfigurationsResponse

Create a value of GetAppInstanceStreamingConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceStreamingConfigurations:GetAppInstanceStreamingConfigurationsResponse', getAppInstanceStreamingConfigurationsResponse_appInstanceStreamingConfigurations - The streaming settings.

$sel:httpStatus:GetAppInstanceStreamingConfigurationsResponse', getAppInstanceStreamingConfigurationsResponse_httpStatus - The response's http status code.

Response Lenses

getAppInstanceStreamingConfigurationsResponse_appInstanceStreamingConfigurations :: Lens' GetAppInstanceStreamingConfigurationsResponse (Maybe (NonEmpty AppInstanceStreamingConfiguration))

The streaming settings.

getAppInstanceStreamingConfigurationsResponse_httpStatus :: Lens' GetAppInstanceStreamingConfigurationsResponse Int

The response's http status code.