Network.AWS.Chime.GetPhoneNumber

Description

Retrieves details for the specified phone number ID, such as associations, capabilities, and product type.

Creating a Request

data GetPhoneNumber

See: newGetPhoneNumber smart constructor.

Constructors

GetPhoneNumber'

Fields

phoneNumberId :: Text

The phone number ID.

newGetPhoneNumber

Arguments

:: Text
$sel:phoneNumberId:GetPhoneNumber'

-> GetPhoneNumber

Create a value of GetPhoneNumber with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumberId:GetPhoneNumber', getPhoneNumber_phoneNumberId - The phone number ID.

Request Lenses

getPhoneNumber_phoneNumberId :: Lens' GetPhoneNumber Text

The phone number ID.

Destructuring the Response

data GetPhoneNumberResponse

See: newGetPhoneNumberResponse smart constructor.

Constructors

GetPhoneNumberResponse'

Fields

phoneNumber :: Maybe PhoneNumber

The phone number details.

httpStatus :: Int

The response's http status code.

newGetPhoneNumberResponse

Arguments

:: Int
$sel:httpStatus:GetPhoneNumberResponse'

-> GetPhoneNumberResponse

Create a value of GetPhoneNumberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumber:GetPhoneNumberResponse', getPhoneNumberResponse_phoneNumber - The phone number details.

$sel:httpStatus:GetPhoneNumberResponse', getPhoneNumberResponse_httpStatus - The response's http status code.

Response Lenses

getPhoneNumberResponse_phoneNumber :: Lens' GetPhoneNumberResponse (Maybe PhoneNumber)

The phone number details.

getPhoneNumberResponse_httpStatus :: Lens' GetPhoneNumberResponse Int

The response's http status code.