Network.AWS.Chime.GetPhoneNumberOrder

Description

Retrieves details for the specified phone number order, such as the order creation timestamp, phone numbers in E.164 format, product type, and order status.

Creating a Request

data GetPhoneNumberOrder

See: newGetPhoneNumberOrder smart constructor.

Constructors

GetPhoneNumberOrder'

Fields

phoneNumberOrderId :: Text

The ID for the phone number order.

newGetPhoneNumberOrder

Arguments

:: Text
$sel:phoneNumberOrderId:GetPhoneNumberOrder'

-> GetPhoneNumberOrder

Create a value of GetPhoneNumberOrder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumberOrderId:GetPhoneNumberOrder', getPhoneNumberOrder_phoneNumberOrderId - The ID for the phone number order.

Request Lenses

getPhoneNumberOrder_phoneNumberOrderId :: Lens' GetPhoneNumberOrder Text

The ID for the phone number order.

Destructuring the Response

data GetPhoneNumberOrderResponse

See: newGetPhoneNumberOrderResponse smart constructor.

Constructors

GetPhoneNumberOrderResponse'

Fields

phoneNumberOrder :: Maybe PhoneNumberOrder

The phone number order details.

httpStatus :: Int

The response's http status code.

newGetPhoneNumberOrderResponse

Arguments

:: Int
$sel:httpStatus:GetPhoneNumberOrderResponse'

-> GetPhoneNumberOrderResponse

Create a value of GetPhoneNumberOrderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumberOrder:GetPhoneNumberOrderResponse', getPhoneNumberOrderResponse_phoneNumberOrder - The phone number order details.

$sel:httpStatus:GetPhoneNumberOrderResponse', getPhoneNumberOrderResponse_httpStatus - The response's http status code.

Response Lenses

getPhoneNumberOrderResponse_phoneNumberOrder :: Lens' GetPhoneNumberOrderResponse (Maybe PhoneNumberOrder)

The phone number order details.

getPhoneNumberOrderResponse_httpStatus :: Lens' GetPhoneNumberOrderResponse Int

The response's http status code.