Network.AWS.Chime.GetRetentionSettings

Description

Gets the retention settings for the specified Amazon Chime Enterprise account. For more information about retention settings, see Managing Chat Retention Policies in the Amazon Chime Administration Guide.

Creating a Request

data GetRetentionSettings

See: newGetRetentionSettings smart constructor.

Constructors

GetRetentionSettings'

Fields

accountId :: Text

The Amazon Chime account ID.

newGetRetentionSettings

Arguments

:: Text
$sel:accountId:GetRetentionSettings'

-> GetRetentionSettings

Create a value of GetRetentionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:GetRetentionSettings', getRetentionSettings_accountId - The Amazon Chime account ID.

Request Lenses

getRetentionSettings_accountId :: Lens' GetRetentionSettings Text

The Amazon Chime account ID.

Destructuring the Response

data GetRetentionSettingsResponse

See: newGetRetentionSettingsResponse smart constructor.

Constructors

GetRetentionSettingsResponse'

Fields

retentionSettings :: Maybe RetentionSettings

The retention settings.

initiateDeletionTimestamp :: Maybe POSIX

The timestamp representing the time at which the specified items are permanently deleted, in ISO 8601 format.

httpStatus :: Int

The response's http status code.

newGetRetentionSettingsResponse

Arguments

:: Int
$sel:httpStatus:GetRetentionSettingsResponse'

-> GetRetentionSettingsResponse

Create a value of GetRetentionSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retentionSettings:GetRetentionSettingsResponse', getRetentionSettingsResponse_retentionSettings - The retention settings.

$sel:initiateDeletionTimestamp:GetRetentionSettingsResponse', getRetentionSettingsResponse_initiateDeletionTimestamp - The timestamp representing the time at which the specified items are permanently deleted, in ISO 8601 format.

$sel:httpStatus:GetRetentionSettingsResponse', getRetentionSettingsResponse_httpStatus - The response's http status code.

Response Lenses

getRetentionSettingsResponse_retentionSettings :: Lens' GetRetentionSettingsResponse (Maybe RetentionSettings)

The retention settings.

getRetentionSettingsResponse_initiateDeletionTimestamp :: Lens' GetRetentionSettingsResponse (Maybe UTCTime)

The timestamp representing the time at which the specified items are permanently deleted, in ISO 8601 format.

getRetentionSettingsResponse_httpStatus :: Lens' GetRetentionSettingsResponse Int

The response's http status code.