Network.AWS.Chime.ListBots

Description

Lists the bots associated with the administrator's Amazon Chime Enterprise account ID.

Creating a Request

data ListBots

See: newListBots smart constructor.

Constructors

ListBots'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results.

maxResults :: Maybe Natural

The maximum number of results to return in a single call. The default is 10.

accountId :: Text

The Amazon Chime account ID.

newListBots

Arguments

:: Text
$sel:accountId:ListBots'

-> ListBots

Create a value of ListBots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBots', listBots_nextToken - The token to use to retrieve the next page of results.

$sel:maxResults:ListBots', listBots_maxResults - The maximum number of results to return in a single call. The default is 10.

$sel:accountId:ListBots', listBots_accountId - The Amazon Chime account ID.

Request Lenses

listBots_nextToken :: Lens' ListBots (Maybe Text)

The token to use to retrieve the next page of results.

listBots_maxResults :: Lens' ListBots (Maybe Natural)

The maximum number of results to return in a single call. The default is 10.

listBots_accountId :: Lens' ListBots Text

The Amazon Chime account ID.

Destructuring the Response

data ListBotsResponse

See: newListBotsResponse smart constructor.

Constructors

ListBotsResponse'

Fields

bots :: Maybe [Bot]

List of bots and bot details.

nextToken :: Maybe Text

The token to use to retrieve the next page of results.

httpStatus :: Int

The response's http status code.

newListBotsResponse

Arguments

:: Int
$sel:httpStatus:ListBotsResponse'

-> ListBotsResponse

Create a value of ListBotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bots:ListBotsResponse', listBotsResponse_bots - List of bots and bot details.

$sel:nextToken:ListBots', listBotsResponse_nextToken - The token to use to retrieve the next page of results.

$sel:httpStatus:ListBotsResponse', listBotsResponse_httpStatus - The response's http status code.

Response Lenses

listBotsResponse_bots :: Lens' ListBotsResponse (Maybe [Bot])

List of bots and bot details.

listBotsResponse_nextToken :: Lens' ListBotsResponse (Maybe Text)

The token to use to retrieve the next page of results.

listBotsResponse_httpStatus :: Lens' ListBotsResponse Int

The response's http status code.