Network.AWS.Chime.ListChannelMembershipsForAppInstanceUser

Description

Lists all channels that a particular AppInstanceUser is a part of. Only an AppInstanceAdmin can call the API with a user ARN that is not their own.

The x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Creating a Request

data ListChannelMembershipsForAppInstanceUser

See: newListChannelMembershipsForAppInstanceUser smart constructor.

Constructors

ListChannelMembershipsForAppInstanceUser'

Fields

appInstanceUserArn :: Maybe Text

The ARN of the AppInstanceUsers

chimeBearer :: Maybe Text

The AppInstanceUserArn of the user that makes the API call.

nextToken :: Maybe (Sensitive Text)

The token returned from previous API requests until the number of channel memberships is reached.

maxResults :: Maybe Natural

The maximum number of users that you want returned.

newListChannelMembershipsForAppInstanceUser :: ListChannelMembershipsForAppInstanceUser

Create a value of ListChannelMembershipsForAppInstanceUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUserArn:ListChannelMembershipsForAppInstanceUser', listChannelMembershipsForAppInstanceUser_appInstanceUserArn - The ARN of the AppInstanceUsers

$sel:chimeBearer:ListChannelMembershipsForAppInstanceUser', listChannelMembershipsForAppInstanceUser_chimeBearer - The AppInstanceUserArn of the user that makes the API call.

$sel:nextToken:ListChannelMembershipsForAppInstanceUser', listChannelMembershipsForAppInstanceUser_nextToken - The token returned from previous API requests until the number of channel memberships is reached.

$sel:maxResults:ListChannelMembershipsForAppInstanceUser', listChannelMembershipsForAppInstanceUser_maxResults - The maximum number of users that you want returned.

Request Lenses

listChannelMembershipsForAppInstanceUser_appInstanceUserArn :: Lens' ListChannelMembershipsForAppInstanceUser (Maybe Text)

The ARN of the AppInstanceUsers

listChannelMembershipsForAppInstanceUser_chimeBearer :: Lens' ListChannelMembershipsForAppInstanceUser (Maybe Text)

The AppInstanceUserArn of the user that makes the API call.

listChannelMembershipsForAppInstanceUser_nextToken :: Lens' ListChannelMembershipsForAppInstanceUser (Maybe Text)

The token returned from previous API requests until the number of channel memberships is reached.

listChannelMembershipsForAppInstanceUser_maxResults :: Lens' ListChannelMembershipsForAppInstanceUser (Maybe Natural)

The maximum number of users that you want returned.

Destructuring the Response

data ListChannelMembershipsForAppInstanceUserResponse

See: newListChannelMembershipsForAppInstanceUserResponse smart constructor.

Constructors

ListChannelMembershipsForAppInstanceUserResponse'

Fields

channelMemberships :: Maybe [ChannelMembershipForAppInstanceUserSummary]

The token passed by previous API calls until all requested users are returned.

nextToken :: Maybe (Sensitive Text)

The token passed by previous API calls until all requested users are returned.

httpStatus :: Int

The response's http status code.

newListChannelMembershipsForAppInstanceUserResponse

Arguments

:: Int
$sel:httpStatus:ListChannelMembershipsForAppInstanceUserResponse'

-> ListChannelMembershipsForAppInstanceUserResponse

Create a value of ListChannelMembershipsForAppInstanceUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelMemberships:ListChannelMembershipsForAppInstanceUserResponse', listChannelMembershipsForAppInstanceUserResponse_channelMemberships - The token passed by previous API calls until all requested users are returned.

$sel:nextToken:ListChannelMembershipsForAppInstanceUser', listChannelMembershipsForAppInstanceUserResponse_nextToken - The token passed by previous API calls until all requested users are returned.

$sel:httpStatus:ListChannelMembershipsForAppInstanceUserResponse', listChannelMembershipsForAppInstanceUserResponse_httpStatus - The response's http status code.

Response Lenses

listChannelMembershipsForAppInstanceUserResponse_channelMemberships :: Lens' ListChannelMembershipsForAppInstanceUserResponse (Maybe [ChannelMembershipForAppInstanceUserSummary])

The token passed by previous API calls until all requested users are returned.

listChannelMembershipsForAppInstanceUserResponse_nextToken :: Lens' ListChannelMembershipsForAppInstanceUserResponse (Maybe Text)

The token passed by previous API calls until all requested users are returned.

listChannelMembershipsForAppInstanceUserResponse_httpStatus :: Lens' ListChannelMembershipsForAppInstanceUserResponse Int

The response's http status code.