Network.AWS.Chime.ListChannelsModeratedByAppInstanceUser

Description

A list of the channels moderated by an AppInstanceUser.

The x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Creating a Request

data ListChannelsModeratedByAppInstanceUser

See: newListChannelsModeratedByAppInstanceUser smart constructor.

Constructors

ListChannelsModeratedByAppInstanceUser'

Fields

appInstanceUserArn :: Maybe Text

The ARN of the user in the moderated channel.

chimeBearer :: Maybe Text

The AppInstanceUserArn of the user that makes the API call.

nextToken :: Maybe (Sensitive Text)

The token returned from previous API requests until the number of channels moderated by the user is reached.

maxResults :: Maybe Natural

The maximum number of channels in the request.

newListChannelsModeratedByAppInstanceUser :: ListChannelsModeratedByAppInstanceUser

Create a value of ListChannelsModeratedByAppInstanceUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUserArn:ListChannelsModeratedByAppInstanceUser', listChannelsModeratedByAppInstanceUser_appInstanceUserArn - The ARN of the user in the moderated channel.

$sel:chimeBearer:ListChannelsModeratedByAppInstanceUser', listChannelsModeratedByAppInstanceUser_chimeBearer - The AppInstanceUserArn of the user that makes the API call.

$sel:nextToken:ListChannelsModeratedByAppInstanceUser', listChannelsModeratedByAppInstanceUser_nextToken - The token returned from previous API requests until the number of channels moderated by the user is reached.

$sel:maxResults:ListChannelsModeratedByAppInstanceUser', listChannelsModeratedByAppInstanceUser_maxResults - The maximum number of channels in the request.

Request Lenses

listChannelsModeratedByAppInstanceUser_appInstanceUserArn :: Lens' ListChannelsModeratedByAppInstanceUser (Maybe Text)

The ARN of the user in the moderated channel.

listChannelsModeratedByAppInstanceUser_chimeBearer :: Lens' ListChannelsModeratedByAppInstanceUser (Maybe Text)

The AppInstanceUserArn of the user that makes the API call.

listChannelsModeratedByAppInstanceUser_nextToken :: Lens' ListChannelsModeratedByAppInstanceUser (Maybe Text)

The token returned from previous API requests until the number of channels moderated by the user is reached.

listChannelsModeratedByAppInstanceUser_maxResults :: Lens' ListChannelsModeratedByAppInstanceUser (Maybe Natural)

The maximum number of channels in the request.

Destructuring the Response

data ListChannelsModeratedByAppInstanceUserResponse

See: newListChannelsModeratedByAppInstanceUserResponse smart constructor.

Constructors

ListChannelsModeratedByAppInstanceUserResponse'

Fields

channels :: Maybe [ChannelModeratedByAppInstanceUserSummary]

The moderated channels in the request.

nextToken :: Maybe (Sensitive Text)

The token returned from previous API requests until the number of channels moderated by the user is reached.

httpStatus :: Int

The response's http status code.

newListChannelsModeratedByAppInstanceUserResponse

Arguments

:: Int
$sel:httpStatus:ListChannelsModeratedByAppInstanceUserResponse'

-> ListChannelsModeratedByAppInstanceUserResponse

Create a value of ListChannelsModeratedByAppInstanceUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channels:ListChannelsModeratedByAppInstanceUserResponse', listChannelsModeratedByAppInstanceUserResponse_channels - The moderated channels in the request.

$sel:nextToken:ListChannelsModeratedByAppInstanceUser', listChannelsModeratedByAppInstanceUserResponse_nextToken - The token returned from previous API requests until the number of channels moderated by the user is reached.

$sel:httpStatus:ListChannelsModeratedByAppInstanceUserResponse', listChannelsModeratedByAppInstanceUserResponse_httpStatus - The response's http status code.

Response Lenses

listChannelsModeratedByAppInstanceUserResponse_channels :: Lens' ListChannelsModeratedByAppInstanceUserResponse (Maybe [ChannelModeratedByAppInstanceUserSummary])

The moderated channels in the request.

listChannelsModeratedByAppInstanceUserResponse_nextToken :: Lens' ListChannelsModeratedByAppInstanceUserResponse (Maybe Text)

The token returned from previous API requests until the number of channels moderated by the user is reached.

listChannelsModeratedByAppInstanceUserResponse_httpStatus :: Lens' ListChannelsModeratedByAppInstanceUserResponse Int

The response's http status code.