Network.AWS.Chime.PutEventsConfiguration

Description

Creates an events configuration that allows a bot to receive outgoing events sent by Amazon Chime. Choose either an HTTPS endpoint or a Lambda function ARN. For more information, see Bot.

Creating a Request

data PutEventsConfiguration

See: newPutEventsConfiguration smart constructor.

Constructors

PutEventsConfiguration'

Fields

lambdaFunctionArn :: Maybe (Sensitive Text)

Lambda function ARN that allows the bot to receive outgoing events.

outboundEventsHTTPSEndpoint :: Maybe (Sensitive Text)

HTTPS endpoint that allows the bot to receive outgoing events.

accountId :: Text

The Amazon Chime account ID.

botId :: Text

The bot ID.

newPutEventsConfiguration

Arguments

:: Text
$sel:accountId:PutEventsConfiguration'

-> Text
$sel:botId:PutEventsConfiguration'

-> PutEventsConfiguration

Create a value of PutEventsConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lambdaFunctionArn:PutEventsConfiguration', putEventsConfiguration_lambdaFunctionArn - Lambda function ARN that allows the bot to receive outgoing events.

$sel:outboundEventsHTTPSEndpoint:PutEventsConfiguration', putEventsConfiguration_outboundEventsHTTPSEndpoint - HTTPS endpoint that allows the bot to receive outgoing events.

$sel:accountId:PutEventsConfiguration', putEventsConfiguration_accountId - The Amazon Chime account ID.

$sel:botId:PutEventsConfiguration', putEventsConfiguration_botId - The bot ID.

Request Lenses

putEventsConfiguration_lambdaFunctionArn :: Lens' PutEventsConfiguration (Maybe Text)

Lambda function ARN that allows the bot to receive outgoing events.

putEventsConfiguration_outboundEventsHTTPSEndpoint :: Lens' PutEventsConfiguration (Maybe Text)

HTTPS endpoint that allows the bot to receive outgoing events.

putEventsConfiguration_accountId :: Lens' PutEventsConfiguration Text

The Amazon Chime account ID.

putEventsConfiguration_botId :: Lens' PutEventsConfiguration Text

The bot ID.

Destructuring the Response

data PutEventsConfigurationResponse

See: newPutEventsConfigurationResponse smart constructor.

Constructors

PutEventsConfigurationResponse'

Fields

eventsConfiguration :: Maybe EventsConfiguration

httpStatus :: Int

The response's http status code.

newPutEventsConfigurationResponse

Arguments

:: Int
$sel:httpStatus:PutEventsConfigurationResponse'

-> PutEventsConfigurationResponse

Create a value of PutEventsConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventsConfiguration:PutEventsConfigurationResponse', putEventsConfigurationResponse_eventsConfiguration - Undocumented member.

$sel:httpStatus:PutEventsConfigurationResponse', putEventsConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

putEventsConfigurationResponse_eventsConfiguration :: Lens' PutEventsConfigurationResponse (Maybe EventsConfiguration)

Undocumented member.

putEventsConfigurationResponse_httpStatus :: Lens' PutEventsConfigurationResponse Int

The response's http status code.