Network.AWS.Chime.PutVoiceConnectorOrigination

Description

Adds origination settings for the specified Amazon Chime Voice Connector.

If emergency calling is configured for the Amazon Chime Voice Connector, it must be deleted prior to turning off origination settings.

Creating a Request

data PutVoiceConnectorOrigination

See: newPutVoiceConnectorOrigination smart constructor.

Constructors

PutVoiceConnectorOrigination'

Fields

voiceConnectorId :: Text

The Amazon Chime Voice Connector ID.

origination :: Origination

The origination setting details to add.

newPutVoiceConnectorOrigination

Arguments

:: Text
$sel:voiceConnectorId:PutVoiceConnectorOrigination'

-> Origination
$sel:origination:PutVoiceConnectorOrigination'

-> PutVoiceConnectorOrigination

Create a value of PutVoiceConnectorOrigination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:voiceConnectorId:PutVoiceConnectorOrigination', putVoiceConnectorOrigination_voiceConnectorId - The Amazon Chime Voice Connector ID.

$sel:origination:PutVoiceConnectorOrigination', putVoiceConnectorOrigination_origination - The origination setting details to add.

Request Lenses

putVoiceConnectorOrigination_voiceConnectorId :: Lens' PutVoiceConnectorOrigination Text

The Amazon Chime Voice Connector ID.

putVoiceConnectorOrigination_origination :: Lens' PutVoiceConnectorOrigination Origination

The origination setting details to add.

Destructuring the Response

data PutVoiceConnectorOriginationResponse

See: newPutVoiceConnectorOriginationResponse smart constructor.

Constructors

PutVoiceConnectorOriginationResponse'

Fields

origination :: Maybe Origination

The updated origination setting details.

httpStatus :: Int

The response's http status code.

newPutVoiceConnectorOriginationResponse

Arguments

:: Int
$sel:httpStatus:PutVoiceConnectorOriginationResponse'

-> PutVoiceConnectorOriginationResponse

Create a value of PutVoiceConnectorOriginationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:origination:PutVoiceConnectorOrigination', putVoiceConnectorOriginationResponse_origination - The updated origination setting details.

$sel:httpStatus:PutVoiceConnectorOriginationResponse', putVoiceConnectorOriginationResponse_httpStatus - The response's http status code.

Response Lenses

putVoiceConnectorOriginationResponse_origination :: Lens' PutVoiceConnectorOriginationResponse (Maybe Origination)

The updated origination setting details.

putVoiceConnectorOriginationResponse_httpStatus :: Lens' PutVoiceConnectorOriginationResponse Int

The response's http status code.