Network.AWS.Chime.Types.AppInstanceStreamingConfiguration

Description

Documentation

data AppInstanceStreamingConfiguration

The details of the streaming configuration of an AppInstance.

See: newAppInstanceStreamingConfiguration smart constructor.

Constructors

AppInstanceStreamingConfiguration'

Fields

appInstanceDataType :: AppInstanceDataType

The type of data to be streamed.

resourceArn :: Sensitive Text

The resource ARN.

newAppInstanceStreamingConfiguration

Arguments

:: AppInstanceDataType
$sel:appInstanceDataType:AppInstanceStreamingConfiguration'

-> Text
$sel:resourceArn:AppInstanceStreamingConfiguration'

-> AppInstanceStreamingConfiguration

Create a value of AppInstanceStreamingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceDataType:AppInstanceStreamingConfiguration', appInstanceStreamingConfiguration_appInstanceDataType - The type of data to be streamed.

$sel:resourceArn:AppInstanceStreamingConfiguration', appInstanceStreamingConfiguration_resourceArn - The resource ARN.

appInstanceStreamingConfiguration_appInstanceDataType :: Lens' AppInstanceStreamingConfiguration AppInstanceDataType

The type of data to be streamed.

appInstanceStreamingConfiguration_resourceArn :: Lens' AppInstanceStreamingConfiguration Text

The resource ARN.