Network.AWS.Chime.Types.Bot

Description

Documentation

data Bot

A resource that allows Enterprise account administrators to configure an interface to receive events from Amazon Chime.

See: newBot smart constructor.

Constructors

Bot'

Fields

securityToken :: Maybe (Sensitive Text)

The security token used to authenticate Amazon Chime with the outgoing event endpoint.

disabled :: Maybe Bool

When true, the bot is stopped from running in your account.

updatedTimestamp :: Maybe POSIX

The updated bot timestamp, in ISO 8601 format.

userId :: Maybe Text

The unique ID for the bot user.

botId :: Maybe Text

The bot ID.

displayName :: Maybe (Sensitive Text)

The bot display name.

botEmail :: Maybe (Sensitive Text)

The bot email address.

createdTimestamp :: Maybe POSIX

The bot creation timestamp, in ISO 8601 format.

botType :: Maybe BotType

The bot type.

newBot :: Bot

Create a value of Bot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityToken:Bot', bot_securityToken - The security token used to authenticate Amazon Chime with the outgoing event endpoint.

$sel:disabled:Bot', bot_disabled - When true, the bot is stopped from running in your account.

$sel:updatedTimestamp:Bot', bot_updatedTimestamp - The updated bot timestamp, in ISO 8601 format.

$sel:userId:Bot', bot_userId - The unique ID for the bot user.

$sel:botId:Bot', bot_botId - The bot ID.

$sel:displayName:Bot', bot_displayName - The bot display name.

$sel:botEmail:Bot', bot_botEmail - The bot email address.

$sel:createdTimestamp:Bot', bot_createdTimestamp - The bot creation timestamp, in ISO 8601 format.

$sel:botType:Bot', bot_botType - The bot type.

bot_securityToken :: Lens' Bot (Maybe Text)

The security token used to authenticate Amazon Chime with the outgoing event endpoint.

bot_disabled :: Lens' Bot (Maybe Bool)

When true, the bot is stopped from running in your account.

bot_updatedTimestamp :: Lens' Bot (Maybe UTCTime)

The updated bot timestamp, in ISO 8601 format.

bot_userId :: Lens' Bot (Maybe Text)

The unique ID for the bot user.

bot_botId :: Lens' Bot (Maybe Text)

The bot ID.

bot_displayName :: Lens' Bot (Maybe Text)

The bot display name.

bot_botEmail :: Lens' Bot (Maybe Text)

The bot email address.

bot_createdTimestamp :: Lens' Bot (Maybe UTCTime)

The bot creation timestamp, in ISO 8601 format.

bot_botType :: Lens' Bot (Maybe BotType)

The bot type.