Network.AWS.Chime.Types.ChannelMembership

Description

Documentation

data ChannelMembership

The details of a channel member.

See: newChannelMembership smart constructor.

Constructors

ChannelMembership'

Fields

channelArn :: Maybe Text

The ARN of the member's channel.

member :: Maybe Identity

The data of the channel member.

type' :: Maybe ChannelMembershipType

The membership type set for the channel member.

invitedBy :: Maybe Identity

The identifier of the member who invited another member.

createdTimestamp :: Maybe POSIX

The time at which the channel membership was created.

lastUpdatedTimestamp :: Maybe POSIX

The time at which a channel membership was last updated.

newChannelMembership :: ChannelMembership

Create a value of ChannelMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:ChannelMembership', channelMembership_channelArn - The ARN of the member's channel.

$sel:member:ChannelMembership', channelMembership_member - The data of the channel member.

$sel:type':ChannelMembership', channelMembership_type - The membership type set for the channel member.

$sel:invitedBy:ChannelMembership', channelMembership_invitedBy - The identifier of the member who invited another member.

$sel:createdTimestamp:ChannelMembership', channelMembership_createdTimestamp - The time at which the channel membership was created.

$sel:lastUpdatedTimestamp:ChannelMembership', channelMembership_lastUpdatedTimestamp - The time at which a channel membership was last updated.

channelMembership_channelArn :: Lens' ChannelMembership (Maybe Text)

The ARN of the member's channel.

channelMembership_member :: Lens' ChannelMembership (Maybe Identity)

The data of the channel member.

channelMembership_type :: Lens' ChannelMembership (Maybe ChannelMembershipType)

The membership type set for the channel member.

channelMembership_invitedBy :: Lens' ChannelMembership (Maybe Identity)

The identifier of the member who invited another member.

channelMembership_createdTimestamp :: Lens' ChannelMembership (Maybe UTCTime)

The time at which the channel membership was created.

channelMembership_lastUpdatedTimestamp :: Lens' ChannelMembership (Maybe UTCTime)

The time at which a channel membership was last updated.