Network.AWS.Chime.Types.ContentArtifactsConfiguration

Description

Documentation

data ContentArtifactsConfiguration

The content artifact object.

See: newContentArtifactsConfiguration smart constructor.

Constructors

ContentArtifactsConfiguration'

Fields

muxType :: Maybe ContentMuxType

The MUX type of the artifact configuration.

state :: ArtifactsState

Indicates whether the content artifact is enabled or disabled.

newContentArtifactsConfiguration

Arguments

:: ArtifactsState
$sel:state:ContentArtifactsConfiguration'

-> ContentArtifactsConfiguration

Create a value of ContentArtifactsConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:muxType:ContentArtifactsConfiguration', contentArtifactsConfiguration_muxType - The MUX type of the artifact configuration.

$sel:state:ContentArtifactsConfiguration', contentArtifactsConfiguration_state - Indicates whether the content artifact is enabled or disabled.

contentArtifactsConfiguration_muxType :: Lens' ContentArtifactsConfiguration (Maybe ContentMuxType)

The MUX type of the artifact configuration.

contentArtifactsConfiguration_state :: Lens' ContentArtifactsConfiguration ArtifactsState

Indicates whether the content artifact is enabled or disabled.