Network.AWS.Chime.Types.MediaCapturePipeline

Description

Documentation

data MediaCapturePipeline

A media capture pipeline object consisting of an ID, source type, source ARN, a sink type, a sink ARN, and a configuration object.

See: newMediaCapturePipeline smart constructor.

Constructors

MediaCapturePipeline'

Fields

status :: Maybe MediaPipelineStatus

The status of the media capture pipeline.

sourceType :: Maybe MediaPipelineSourceType

Source type from which media artifacts are saved. You must use ChimeMeeting.

sourceArn :: Maybe (Sensitive Text)

ARN of the source from which the media artifacts will be saved.

updatedTimestamp :: Maybe POSIX

The time at which the capture pipeline was updated, in ISO 8601 format.

sinkType :: Maybe MediaPipelineSinkType

Destination type to which the media artifacts are saved. You must use an S3 Bucket.

chimeSdkMeetingConfiguration :: Maybe ChimeSdkMeetingConfiguration

The configuration for a specified media capture pipeline. SourceType must be ChimeSdkMeeting.

sinkArn :: Maybe (Sensitive Text)

ARN of the destination to which the media artifacts are saved.

mediaPipelineId :: Maybe Text

The ID of a media capture pipeline.

createdTimestamp :: Maybe POSIX

The time at which the capture pipeline was created, in ISO 8601 format.

newMediaCapturePipeline :: MediaCapturePipeline

Create a value of MediaCapturePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:MediaCapturePipeline', mediaCapturePipeline_status - The status of the media capture pipeline.

$sel:sourceType:MediaCapturePipeline', mediaCapturePipeline_sourceType - Source type from which media artifacts are saved. You must use ChimeMeeting.

$sel:sourceArn:MediaCapturePipeline', mediaCapturePipeline_sourceArn - ARN of the source from which the media artifacts will be saved.

$sel:updatedTimestamp:MediaCapturePipeline', mediaCapturePipeline_updatedTimestamp - The time at which the capture pipeline was updated, in ISO 8601 format.

$sel:sinkType:MediaCapturePipeline', mediaCapturePipeline_sinkType - Destination type to which the media artifacts are saved. You must use an S3 Bucket.

$sel:chimeSdkMeetingConfiguration:MediaCapturePipeline', mediaCapturePipeline_chimeSdkMeetingConfiguration - The configuration for a specified media capture pipeline. SourceType must be ChimeSdkMeeting.

$sel:sinkArn:MediaCapturePipeline', mediaCapturePipeline_sinkArn - ARN of the destination to which the media artifacts are saved.

$sel:mediaPipelineId:MediaCapturePipeline', mediaCapturePipeline_mediaPipelineId - The ID of a media capture pipeline.

$sel:createdTimestamp:MediaCapturePipeline', mediaCapturePipeline_createdTimestamp - The time at which the capture pipeline was created, in ISO 8601 format.

mediaCapturePipeline_status :: Lens' MediaCapturePipeline (Maybe MediaPipelineStatus)

The status of the media capture pipeline.

mediaCapturePipeline_sourceType :: Lens' MediaCapturePipeline (Maybe MediaPipelineSourceType)

Source type from which media artifacts are saved. You must use ChimeMeeting.

mediaCapturePipeline_sourceArn :: Lens' MediaCapturePipeline (Maybe Text)

ARN of the source from which the media artifacts will be saved.

mediaCapturePipeline_updatedTimestamp :: Lens' MediaCapturePipeline (Maybe UTCTime)

The time at which the capture pipeline was updated, in ISO 8601 format.

mediaCapturePipeline_sinkType :: Lens' MediaCapturePipeline (Maybe MediaPipelineSinkType)

Destination type to which the media artifacts are saved. You must use an S3 Bucket.

mediaCapturePipeline_chimeSdkMeetingConfiguration :: Lens' MediaCapturePipeline (Maybe ChimeSdkMeetingConfiguration)

The configuration for a specified media capture pipeline. SourceType must be ChimeSdkMeeting.

mediaCapturePipeline_sinkArn :: Lens' MediaCapturePipeline (Maybe Text)

ARN of the destination to which the media artifacts are saved.

mediaCapturePipeline_mediaPipelineId :: Lens' MediaCapturePipeline (Maybe Text)

The ID of a media capture pipeline.

mediaCapturePipeline_createdTimestamp :: Lens' MediaCapturePipeline (Maybe UTCTime)

The time at which the capture pipeline was created, in ISO 8601 format.