Network.AWS.Chime.Types.MediaPlacement

Description

Documentation

data MediaPlacement

A set of endpoints used by clients to connect to the media service group for an Amazon Chime SDK meeting.

See: newMediaPlacement smart constructor.

Constructors

MediaPlacement'

Fields

screenDataUrl :: Maybe Text

The screen data URL.

eventIngestionUrl :: Maybe Text

The event ingestion URL.

signalingUrl :: Maybe Text

The signaling URL.

screenSharingUrl :: Maybe Text

The screen sharing URL.

screenViewingUrl :: Maybe Text

The screen viewing URL.

audioHostUrl :: Maybe Text

The audio host URL.

audioFallbackUrl :: Maybe Text

The audio fallback URL.

turnControlUrl :: Maybe Text

The turn control URL.

newMediaPlacement :: MediaPlacement

Create a value of MediaPlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:screenDataUrl:MediaPlacement', mediaPlacement_screenDataUrl - The screen data URL.

$sel:eventIngestionUrl:MediaPlacement', mediaPlacement_eventIngestionUrl - The event ingestion URL.

$sel:signalingUrl:MediaPlacement', mediaPlacement_signalingUrl - The signaling URL.

$sel:screenSharingUrl:MediaPlacement', mediaPlacement_screenSharingUrl - The screen sharing URL.

$sel:screenViewingUrl:MediaPlacement', mediaPlacement_screenViewingUrl - The screen viewing URL.

$sel:audioHostUrl:MediaPlacement', mediaPlacement_audioHostUrl - The audio host URL.

$sel:audioFallbackUrl:MediaPlacement', mediaPlacement_audioFallbackUrl - The audio fallback URL.

$sel:turnControlUrl:MediaPlacement', mediaPlacement_turnControlUrl - The turn control URL.

mediaPlacement_screenDataUrl :: Lens' MediaPlacement (Maybe Text)

The screen data URL.

mediaPlacement_eventIngestionUrl :: Lens' MediaPlacement (Maybe Text)

The event ingestion URL.

mediaPlacement_signalingUrl :: Lens' MediaPlacement (Maybe Text)

The signaling URL.

mediaPlacement_screenSharingUrl :: Lens' MediaPlacement (Maybe Text)

The screen sharing URL.

mediaPlacement_screenViewingUrl :: Lens' MediaPlacement (Maybe Text)

The screen viewing URL.

mediaPlacement_audioHostUrl :: Lens' MediaPlacement (Maybe Text)

The audio host URL.

mediaPlacement_audioFallbackUrl :: Lens' MediaPlacement (Maybe Text)

The audio fallback URL.

mediaPlacement_turnControlUrl :: Lens' MediaPlacement (Maybe Text)

The turn control URL.