Network.AWS.Chime.Types.PhoneNumber

Description

Documentation

data PhoneNumber

A phone number used for Amazon Chime Business Calling or an Amazon Chime Voice Connector.

See: newPhoneNumber smart constructor.

Constructors

PhoneNumber'

Fields

status :: Maybe PhoneNumberStatus

The phone number status.

deletionTimestamp :: Maybe POSIX

The deleted phone number timestamp, in ISO 8601 format.

phoneNumberId :: Maybe Text

The phone number ID.

country :: Maybe Text

The phone number country. Format: ISO 3166-1 alpha-2.

updatedTimestamp :: Maybe POSIX

The updated phone number timestamp, in ISO 8601 format.

productType :: Maybe PhoneNumberProductType

The phone number product type.

e164PhoneNumber :: Maybe (Sensitive Text)

The phone number, in E.164 format.

associations :: Maybe [PhoneNumberAssociation]

The phone number associations.

callingName :: Maybe (Sensitive Text)

The outbound calling name associated with the phone number.

type' :: Maybe PhoneNumberType

The phone number type.

createdTimestamp :: Maybe POSIX

The phone number creation timestamp, in ISO 8601 format.

capabilities :: Maybe PhoneNumberCapabilities

The phone number capabilities.

callingNameStatus :: Maybe CallingNameStatus

The outbound calling name status.

newPhoneNumber :: PhoneNumber

Create a value of PhoneNumber with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:PhoneNumber', phoneNumber_status - The phone number status.

$sel:deletionTimestamp:PhoneNumber', phoneNumber_deletionTimestamp - The deleted phone number timestamp, in ISO 8601 format.

$sel:phoneNumberId:PhoneNumber', phoneNumber_phoneNumberId - The phone number ID.

$sel:country:PhoneNumber', phoneNumber_country - The phone number country. Format: ISO 3166-1 alpha-2.

$sel:updatedTimestamp:PhoneNumber', phoneNumber_updatedTimestamp - The updated phone number timestamp, in ISO 8601 format.

$sel:productType:PhoneNumber', phoneNumber_productType - The phone number product type.

$sel:e164PhoneNumber:PhoneNumber', phoneNumber_e164PhoneNumber - The phone number, in E.164 format.

$sel:associations:PhoneNumber', phoneNumber_associations - The phone number associations.

$sel:callingName:PhoneNumber', phoneNumber_callingName - The outbound calling name associated with the phone number.

$sel:type':PhoneNumber', phoneNumber_type - The phone number type.

$sel:createdTimestamp:PhoneNumber', phoneNumber_createdTimestamp - The phone number creation timestamp, in ISO 8601 format.

$sel:capabilities:PhoneNumber', phoneNumber_capabilities - The phone number capabilities.

$sel:callingNameStatus:PhoneNumber', phoneNumber_callingNameStatus - The outbound calling name status.

phoneNumber_status :: Lens' PhoneNumber (Maybe PhoneNumberStatus)

The phone number status.

phoneNumber_deletionTimestamp :: Lens' PhoneNumber (Maybe UTCTime)

The deleted phone number timestamp, in ISO 8601 format.

phoneNumber_phoneNumberId :: Lens' PhoneNumber (Maybe Text)

The phone number ID.

phoneNumber_country :: Lens' PhoneNumber (Maybe Text)

The phone number country. Format: ISO 3166-1 alpha-2.

phoneNumber_updatedTimestamp :: Lens' PhoneNumber (Maybe UTCTime)

The updated phone number timestamp, in ISO 8601 format.

phoneNumber_productType :: Lens' PhoneNumber (Maybe PhoneNumberProductType)

The phone number product type.

phoneNumber_e164PhoneNumber :: Lens' PhoneNumber (Maybe Text)

The phone number, in E.164 format.

phoneNumber_associations :: Lens' PhoneNumber (Maybe [PhoneNumberAssociation])

The phone number associations.

phoneNumber_callingName :: Lens' PhoneNumber (Maybe Text)

The outbound calling name associated with the phone number.

phoneNumber_type :: Lens' PhoneNumber (Maybe PhoneNumberType)

The phone number type.

phoneNumber_createdTimestamp :: Lens' PhoneNumber (Maybe UTCTime)

The phone number creation timestamp, in ISO 8601 format.

phoneNumber_capabilities :: Lens' PhoneNumber (Maybe PhoneNumberCapabilities)

The phone number capabilities.

phoneNumber_callingNameStatus :: Lens' PhoneNumber (Maybe CallingNameStatus)

The outbound calling name status.