Network.AWS.Chime.Types.PhoneNumberCapabilities

Description

Documentation

data PhoneNumberCapabilities

The phone number capabilities for Amazon Chime Business Calling phone numbers, such as enabled inbound and outbound calling and text messaging.

See: newPhoneNumberCapabilities smart constructor.

Constructors

PhoneNumberCapabilities'

Fields

outboundMMS :: Maybe Bool

Allows or denies outbound MMS messaging for the specified phone number.

inboundCall :: Maybe Bool

Allows or denies inbound calling for the specified phone number.

inboundSMS :: Maybe Bool

Allows or denies inbound SMS messaging for the specified phone number.

inboundMMS :: Maybe Bool

Allows or denies inbound MMS messaging for the specified phone number.

outboundCall :: Maybe Bool

Allows or denies outbound calling for the specified phone number.

outboundSMS :: Maybe Bool

Allows or denies outbound SMS messaging for the specified phone number.

newPhoneNumberCapabilities :: PhoneNumberCapabilities

Create a value of PhoneNumberCapabilities with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outboundMMS:PhoneNumberCapabilities', phoneNumberCapabilities_outboundMMS - Allows or denies outbound MMS messaging for the specified phone number.

$sel:inboundCall:PhoneNumberCapabilities', phoneNumberCapabilities_inboundCall - Allows or denies inbound calling for the specified phone number.

$sel:inboundSMS:PhoneNumberCapabilities', phoneNumberCapabilities_inboundSMS - Allows or denies inbound SMS messaging for the specified phone number.

$sel:inboundMMS:PhoneNumberCapabilities', phoneNumberCapabilities_inboundMMS - Allows or denies inbound MMS messaging for the specified phone number.

$sel:outboundCall:PhoneNumberCapabilities', phoneNumberCapabilities_outboundCall - Allows or denies outbound calling for the specified phone number.

$sel:outboundSMS:PhoneNumberCapabilities', phoneNumberCapabilities_outboundSMS - Allows or denies outbound SMS messaging for the specified phone number.

phoneNumberCapabilities_outboundMMS :: Lens' PhoneNumberCapabilities (Maybe Bool)

Allows or denies outbound MMS messaging for the specified phone number.

phoneNumberCapabilities_inboundCall :: Lens' PhoneNumberCapabilities (Maybe Bool)

Allows or denies inbound calling for the specified phone number.

phoneNumberCapabilities_inboundSMS :: Lens' PhoneNumberCapabilities (Maybe Bool)

Allows or denies inbound SMS messaging for the specified phone number.

phoneNumberCapabilities_inboundMMS :: Lens' PhoneNumberCapabilities (Maybe Bool)

Allows or denies inbound MMS messaging for the specified phone number.

phoneNumberCapabilities_outboundCall :: Lens' PhoneNumberCapabilities (Maybe Bool)

Allows or denies outbound calling for the specified phone number.

phoneNumberCapabilities_outboundSMS :: Lens' PhoneNumberCapabilities (Maybe Bool)

Allows or denies outbound SMS messaging for the specified phone number.