Network.AWS.Chime.Types.Proxy

Description

Documentation

data Proxy

The proxy configuration for an Amazon Chime Voice Connector.

See: newProxy smart constructor.

Constructors

Proxy'

Fields

defaultSessionExpiryMinutes :: Maybe Int

The default number of minutes allowed for proxy sessions.

disabled :: Maybe Bool

When true, stops proxy sessions from being created on the specified Amazon Chime Voice Connector.

fallBackPhoneNumber :: Maybe (Sensitive Text)

The phone number to route calls to after a proxy session expires.

phoneNumberCountries :: Maybe [Text]

The countries for proxy phone numbers to be selected from.

newProxy :: Proxy

Create a value of Proxy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultSessionExpiryMinutes:Proxy', proxy_defaultSessionExpiryMinutes - The default number of minutes allowed for proxy sessions.

$sel:disabled:Proxy', proxy_disabled - When true, stops proxy sessions from being created on the specified Amazon Chime Voice Connector.

$sel:fallBackPhoneNumber:Proxy', proxy_fallBackPhoneNumber - The phone number to route calls to after a proxy session expires.

$sel:phoneNumberCountries:Proxy', proxy_phoneNumberCountries - The countries for proxy phone numbers to be selected from.

proxy_defaultSessionExpiryMinutes :: Lens' Proxy (Maybe Int)

The default number of minutes allowed for proxy sessions.

proxy_disabled :: Lens' Proxy (Maybe Bool)

When true, stops proxy sessions from being created on the specified Amazon Chime Voice Connector.

proxy_fallBackPhoneNumber :: Lens' Proxy (Maybe Text)

The phone number to route calls to after a proxy session expires.

proxy_phoneNumberCountries :: Lens' Proxy (Maybe [Text])

The countries for proxy phone numbers to be selected from.