Network.AWS.Chime.Types.RoomMembership

Description

Documentation

data RoomMembership

The room membership details.

See: newRoomMembership smart constructor.

Constructors

RoomMembership'

Fields

updatedTimestamp :: Maybe POSIX

The room membership update timestamp, in ISO 8601 format.

role' :: Maybe RoomMembershipRole

The membership role.

roomId :: Maybe Text

The room ID.

member :: Maybe Member

invitedBy :: Maybe Text

The identifier of the user that invited the room member.

newRoomMembership :: RoomMembership

Create a value of RoomMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updatedTimestamp:RoomMembership', roomMembership_updatedTimestamp - The room membership update timestamp, in ISO 8601 format.

$sel:role':RoomMembership', roomMembership_role - The membership role.

$sel:roomId:RoomMembership', roomMembership_roomId - The room ID.

$sel:member:RoomMembership', roomMembership_member - Undocumented member.

$sel:invitedBy:RoomMembership', roomMembership_invitedBy - The identifier of the user that invited the room member.

roomMembership_updatedTimestamp :: Lens' RoomMembership (Maybe UTCTime)

The room membership update timestamp, in ISO 8601 format.

roomMembership_role :: Lens' RoomMembership (Maybe RoomMembershipRole)

The membership role.

roomMembership_roomId :: Lens' RoomMembership (Maybe Text)

The room ID.

roomMembership_member :: Lens' RoomMembership (Maybe Member)

Undocumented member.

roomMembership_invitedBy :: Lens' RoomMembership (Maybe Text)

The identifier of the user that invited the room member.