Network.AWS.Chime.Types.SipRule

Description

Documentation

data SipRule

The SIP rule details, including name, triggers, and target applications. An AWS account can have multiple SIP rules.

See: newSipRule smart constructor.

Constructors

SipRule'

Fields

disabled :: Maybe Bool

Indicates whether the SIP rule is enabled or disabled. You must disable a rule before you can delete it.

targetApplications :: Maybe (NonEmpty SipRuleTargetApplication)

Target SIP media application and other details, such as priority and AWS Region, to be specified in the SIP rule. Only one SIP rule per AWS Region can be provided.

triggerType :: Maybe SipRuleTriggerType

The type of trigger assigned to the SIP rule in TriggerValue, currently RequestUriHostname or ToPhoneNumber.

updatedTimestamp :: Maybe POSIX

The time at which the SIP rule was last updated, in ISO 8601 format.

name :: Maybe Text

The name of the SIP rule.

triggerValue :: Maybe Text

If TriggerType is RequestUriHostname, then the value can be the outbound host name of the Amazon Chime Voice Connector. If TriggerType is ToPhoneNumber, then the value can be a customer-owned phone number in E164 format. SipRule is triggered when a SIP rule requests host name or ToPhoneNumber matches in the incoming SIP request.

createdTimestamp :: Maybe POSIX

The time at which the SIP rule was created, in ISO 8601 format.

sipRuleId :: Maybe Text

The SIP rule ID.

newSipRule :: SipRule

Create a value of SipRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disabled:SipRule', sipRule_disabled - Indicates whether the SIP rule is enabled or disabled. You must disable a rule before you can delete it.

$sel:targetApplications:SipRule', sipRule_targetApplications - Target SIP media application and other details, such as priority and AWS Region, to be specified in the SIP rule. Only one SIP rule per AWS Region can be provided.

$sel:triggerType:SipRule', sipRule_triggerType - The type of trigger assigned to the SIP rule in TriggerValue, currently RequestUriHostname or ToPhoneNumber.

$sel:updatedTimestamp:SipRule', sipRule_updatedTimestamp - The time at which the SIP rule was last updated, in ISO 8601 format.

$sel:name:SipRule', sipRule_name - The name of the SIP rule.

$sel:triggerValue:SipRule', sipRule_triggerValue - If TriggerType is RequestUriHostname, then the value can be the outbound host name of the Amazon Chime Voice Connector. If TriggerType is ToPhoneNumber, then the value can be a customer-owned phone number in E164 format. SipRule is triggered when a SIP rule requests host name or ToPhoneNumber matches in the incoming SIP request.

$sel:createdTimestamp:SipRule', sipRule_createdTimestamp - The time at which the SIP rule was created, in ISO 8601 format.

$sel:sipRuleId:SipRule', sipRule_sipRuleId - The SIP rule ID.

sipRule_disabled :: Lens' SipRule (Maybe Bool)

Indicates whether the SIP rule is enabled or disabled. You must disable a rule before you can delete it.

sipRule_targetApplications :: Lens' SipRule (Maybe (NonEmpty SipRuleTargetApplication))

Target SIP media application and other details, such as priority and AWS Region, to be specified in the SIP rule. Only one SIP rule per AWS Region can be provided.

sipRule_triggerType :: Lens' SipRule (Maybe SipRuleTriggerType)

The type of trigger assigned to the SIP rule in TriggerValue, currently RequestUriHostname or ToPhoneNumber.

sipRule_updatedTimestamp :: Lens' SipRule (Maybe UTCTime)

The time at which the SIP rule was last updated, in ISO 8601 format.

sipRule_name :: Lens' SipRule (Maybe Text)

The name of the SIP rule.

sipRule_triggerValue :: Lens' SipRule (Maybe Text)

If TriggerType is RequestUriHostname, then the value can be the outbound host name of the Amazon Chime Voice Connector. If TriggerType is ToPhoneNumber, then the value can be a customer-owned phone number in E164 format. SipRule is triggered when a SIP rule requests host name or ToPhoneNumber matches in the incoming SIP request.

sipRule_createdTimestamp :: Lens' SipRule (Maybe UTCTime)

The time at which the SIP rule was created, in ISO 8601 format.

sipRule_sipRuleId :: Lens' SipRule (Maybe Text)

The SIP rule ID.