Network.AWS.Chime.Types.User

Description

Documentation

data User

The user on the Amazon Chime account.

See: newUser smart constructor.

Constructors

User'

Fields

userInvitationStatus :: Maybe InviteStatus

The user invite status.

personalPIN :: Maybe Text

The user's personal meeting PIN.

primaryProvisionedNumber :: Maybe (Sensitive Text)

The primary phone number associated with the user.

licenseType :: Maybe License

The license type for the user.

registeredOn :: Maybe POSIX

Date and time when the user is registered, in ISO 8601 format.

accountId :: Maybe Text

The Amazon Chime account ID.

userRegistrationStatus :: Maybe RegistrationStatus

The user registration status.

invitedOn :: Maybe POSIX

Date and time when the user is invited to the Amazon Chime account, in ISO 8601 format.

displayName :: Maybe (Sensitive Text)

The display name of the user.

primaryEmail :: Maybe (Sensitive Text)

The primary email address of the user.

userType :: Maybe UserType

The user type.

alexaForBusinessMetadata :: Maybe AlexaForBusinessMetadata

The Alexa for Business metadata.

userId :: Text

The user ID.

newUser

Arguments

:: Text
$sel:userId:User'

-> User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userInvitationStatus:User', user_userInvitationStatus - The user invite status.

$sel:personalPIN:User', user_personalPIN - The user's personal meeting PIN.

$sel:primaryProvisionedNumber:User', user_primaryProvisionedNumber - The primary phone number associated with the user.

$sel:licenseType:User', user_licenseType - The license type for the user.

$sel:registeredOn:User', user_registeredOn - Date and time when the user is registered, in ISO 8601 format.

$sel:accountId:User', user_accountId - The Amazon Chime account ID.

$sel:userRegistrationStatus:User', user_userRegistrationStatus - The user registration status.

$sel:invitedOn:User', user_invitedOn - Date and time when the user is invited to the Amazon Chime account, in ISO 8601 format.

$sel:displayName:User', user_displayName - The display name of the user.

$sel:primaryEmail:User', user_primaryEmail - The primary email address of the user.

$sel:userType:User', user_userType - The user type.

$sel:alexaForBusinessMetadata:User', user_alexaForBusinessMetadata - The Alexa for Business metadata.

$sel:userId:User', user_userId - The user ID.

user_userInvitationStatus :: Lens' User (Maybe InviteStatus)

The user invite status.

user_personalPIN :: Lens' User (Maybe Text)

The user's personal meeting PIN.

user_primaryProvisionedNumber :: Lens' User (Maybe Text)

The primary phone number associated with the user.

user_licenseType :: Lens' User (Maybe License)

The license type for the user.

user_registeredOn :: Lens' User (Maybe UTCTime)

Date and time when the user is registered, in ISO 8601 format.

user_accountId :: Lens' User (Maybe Text)

The Amazon Chime account ID.

user_userRegistrationStatus :: Lens' User (Maybe RegistrationStatus)

The user registration status.

user_invitedOn :: Lens' User (Maybe UTCTime)

Date and time when the user is invited to the Amazon Chime account, in ISO 8601 format.

user_displayName :: Lens' User (Maybe Text)

The display name of the user.

user_primaryEmail :: Lens' User (Maybe Text)

The primary email address of the user.

user_userType :: Lens' User (Maybe UserType)

The user type.

user_alexaForBusinessMetadata :: Lens' User (Maybe AlexaForBusinessMetadata)

The Alexa for Business metadata.

user_userId :: Lens' User Text

The user ID.