Network.AWS.Chime.Types.UserError

Description

Documentation

data UserError

The list of errors returned when errors are encountered during the BatchSuspendUser, BatchUnsuspendUser, or BatchUpdateUser actions. This includes user IDs, error codes, and error messages.

See: newUserError smart constructor.

Constructors

UserError'

Fields

userId :: Maybe Text

The user ID for which the action failed.

errorCode :: Maybe ErrorCode

The error code.

errorMessage :: Maybe Text

The error message.

newUserError :: UserError

Create a value of UserError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userId:UserError', userError_userId - The user ID for which the action failed.

$sel:errorCode:UserError', userError_errorCode - The error code.

$sel:errorMessage:UserError', userError_errorMessage - The error message.

userError_userId :: Lens' UserError (Maybe Text)

The user ID for which the action failed.

userError_errorCode :: Lens' UserError (Maybe ErrorCode)

The error code.

userError_errorMessage :: Lens' UserError (Maybe Text)

The error message.