Network.AWS.Chime.Types.VoiceConnector

Description

Documentation

data VoiceConnector

The Amazon Chime Voice Connector configuration, including outbound host name and encryption settings.

See: newVoiceConnector smart constructor.

Constructors

VoiceConnector'

Fields

updatedTimestamp :: Maybe POSIX

The updated Amazon Chime Voice Connector timestamp, in ISO 8601 format.

outboundHostName :: Maybe Text

The outbound host name for the Amazon Chime Voice Connector.

name :: Maybe Text

The name of the Amazon Chime Voice Connector.

requireEncryption :: Maybe Bool

Designates whether encryption is required for the Amazon Chime Voice Connector.

awsRegion :: Maybe VoiceConnectorAwsRegion

The AWS Region in which the Amazon Chime Voice Connector is created. Default: us-east-1.

voiceConnectorId :: Maybe Text

The Amazon Chime Voice Connector ID.

voiceConnectorArn :: Maybe Text

The ARN of the specified Amazon Chime Voice Connector.

createdTimestamp :: Maybe POSIX

The Amazon Chime Voice Connector creation timestamp, in ISO 8601 format.

newVoiceConnector :: VoiceConnector

Create a value of VoiceConnector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updatedTimestamp:VoiceConnector', voiceConnector_updatedTimestamp - The updated Amazon Chime Voice Connector timestamp, in ISO 8601 format.

$sel:outboundHostName:VoiceConnector', voiceConnector_outboundHostName - The outbound host name for the Amazon Chime Voice Connector.

$sel:name:VoiceConnector', voiceConnector_name - The name of the Amazon Chime Voice Connector.

$sel:requireEncryption:VoiceConnector', voiceConnector_requireEncryption - Designates whether encryption is required for the Amazon Chime Voice Connector.

$sel:awsRegion:VoiceConnector', voiceConnector_awsRegion - The AWS Region in which the Amazon Chime Voice Connector is created. Default: us-east-1.

$sel:voiceConnectorId:VoiceConnector', voiceConnector_voiceConnectorId - The Amazon Chime Voice Connector ID.

$sel:voiceConnectorArn:VoiceConnector', voiceConnector_voiceConnectorArn - The ARN of the specified Amazon Chime Voice Connector.

$sel:createdTimestamp:VoiceConnector', voiceConnector_createdTimestamp - The Amazon Chime Voice Connector creation timestamp, in ISO 8601 format.

voiceConnector_updatedTimestamp :: Lens' VoiceConnector (Maybe UTCTime)

The updated Amazon Chime Voice Connector timestamp, in ISO 8601 format.

voiceConnector_outboundHostName :: Lens' VoiceConnector (Maybe Text)

The outbound host name for the Amazon Chime Voice Connector.

voiceConnector_name :: Lens' VoiceConnector (Maybe Text)

The name of the Amazon Chime Voice Connector.

voiceConnector_requireEncryption :: Lens' VoiceConnector (Maybe Bool)

Designates whether encryption is required for the Amazon Chime Voice Connector.

voiceConnector_awsRegion :: Lens' VoiceConnector (Maybe VoiceConnectorAwsRegion)

The AWS Region in which the Amazon Chime Voice Connector is created. Default: us-east-1.

voiceConnector_voiceConnectorId :: Lens' VoiceConnector (Maybe Text)

The Amazon Chime Voice Connector ID.

voiceConnector_voiceConnectorArn :: Lens' VoiceConnector (Maybe Text)

The ARN of the specified Amazon Chime Voice Connector.

voiceConnector_createdTimestamp :: Lens' VoiceConnector (Maybe UTCTime)

The Amazon Chime Voice Connector creation timestamp, in ISO 8601 format.