Network.AWS.Chime.UpdatePhoneNumber

Description

Updates phone number details, such as product type or calling name, for the specified phone number ID. You can update one phone number detail at a time. For example, you can update either the product type or the calling name in one action.

For toll-free numbers, you cannot use the Amazon Chime Business Calling product type. For numbers outside the U.S., you must use the Amazon Chime SIP Media Application Dial-In product type.

Updates to outbound calling names can take 72 hours to complete. Pending updates to outbound calling names must be complete before you can request another update.

Creating a Request

data UpdatePhoneNumber

See: newUpdatePhoneNumber smart constructor.

Constructors

UpdatePhoneNumber'

Fields

productType :: Maybe PhoneNumberProductType

The product type.

callingName :: Maybe (Sensitive Text)

The outbound calling name associated with the phone number.

phoneNumberId :: Text

The phone number ID.

newUpdatePhoneNumber

Arguments

:: Text
$sel:phoneNumberId:UpdatePhoneNumber'

-> UpdatePhoneNumber

Create a value of UpdatePhoneNumber with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:productType:UpdatePhoneNumber', updatePhoneNumber_productType - The product type.

$sel:callingName:UpdatePhoneNumber', updatePhoneNumber_callingName - The outbound calling name associated with the phone number.

$sel:phoneNumberId:UpdatePhoneNumber', updatePhoneNumber_phoneNumberId - The phone number ID.

Request Lenses

updatePhoneNumber_productType :: Lens' UpdatePhoneNumber (Maybe PhoneNumberProductType)

The product type.

updatePhoneNumber_callingName :: Lens' UpdatePhoneNumber (Maybe Text)

The outbound calling name associated with the phone number.

updatePhoneNumber_phoneNumberId :: Lens' UpdatePhoneNumber Text

The phone number ID.

Destructuring the Response

data UpdatePhoneNumberResponse

See: newUpdatePhoneNumberResponse smart constructor.

Constructors

UpdatePhoneNumberResponse'

Fields

phoneNumber :: Maybe PhoneNumber

The updated phone number details.

httpStatus :: Int

The response's http status code.

newUpdatePhoneNumberResponse

Arguments

:: Int
$sel:httpStatus:UpdatePhoneNumberResponse'

-> UpdatePhoneNumberResponse

Create a value of UpdatePhoneNumberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneNumber:UpdatePhoneNumberResponse', updatePhoneNumberResponse_phoneNumber - The updated phone number details.

$sel:httpStatus:UpdatePhoneNumberResponse', updatePhoneNumberResponse_httpStatus - The response's http status code.

Response Lenses

updatePhoneNumberResponse_phoneNumber :: Lens' UpdatePhoneNumberResponse (Maybe PhoneNumber)

The updated phone number details.

updatePhoneNumberResponse_httpStatus :: Lens' UpdatePhoneNumberResponse Int

The response's http status code.