Network.AWS.Chime.UpdateRoom

Description

Updates room details, such as the room name, for a room in an Amazon Chime Enterprise account.

Creating a Request

data UpdateRoom

See: newUpdateRoom smart constructor.

Constructors

UpdateRoom'

Fields

name :: Maybe (Sensitive Text)

The room name.

accountId :: Text

The Amazon Chime account ID.

roomId :: Text

The room ID.

newUpdateRoom

Arguments

:: Text
$sel:accountId:UpdateRoom'

-> Text
$sel:roomId:UpdateRoom'

-> UpdateRoom

Create a value of UpdateRoom with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateRoom', updateRoom_name - The room name.

$sel:accountId:UpdateRoom', updateRoom_accountId - The Amazon Chime account ID.

$sel:roomId:UpdateRoom', updateRoom_roomId - The room ID.

Request Lenses

updateRoom_name :: Lens' UpdateRoom (Maybe Text)

The room name.

updateRoom_accountId :: Lens' UpdateRoom Text

The Amazon Chime account ID.

updateRoom_roomId :: Lens' UpdateRoom Text

The room ID.

Destructuring the Response

data UpdateRoomResponse

See: newUpdateRoomResponse smart constructor.

Constructors

UpdateRoomResponse'

Fields

room :: Maybe Room

The room details.

httpStatus :: Int

The response's http status code.

newUpdateRoomResponse

Arguments

:: Int
$sel:httpStatus:UpdateRoomResponse'

-> UpdateRoomResponse

Create a value of UpdateRoomResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:room:UpdateRoomResponse', updateRoomResponse_room - The room details.

$sel:httpStatus:UpdateRoomResponse', updateRoomResponse_httpStatus - The response's http status code.

Response Lenses

updateRoomResponse_room :: Lens' UpdateRoomResponse (Maybe Room)

The room details.

updateRoomResponse_httpStatus :: Lens' UpdateRoomResponse Int

The response's http status code.