Network.AWS.Chime.UpdateSipMediaApplication

Description

Updates the details of the specified SIP media application.

Creating a Request

data UpdateSipMediaApplication

See: newUpdateSipMediaApplication smart constructor.

Constructors

UpdateSipMediaApplication'

Fields

name :: Maybe Text

The new name for the specified SIP media application.

endpoints :: Maybe (NonEmpty SipMediaApplicationEndpoint)

The new set of endpoints for the specified SIP media application.

sipMediaApplicationId :: Text

The SIP media application ID.

newUpdateSipMediaApplication

Arguments

:: Text
$sel:sipMediaApplicationId:UpdateSipMediaApplication'

-> UpdateSipMediaApplication

Create a value of UpdateSipMediaApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateSipMediaApplication', updateSipMediaApplication_name - The new name for the specified SIP media application.

$sel:endpoints:UpdateSipMediaApplication', updateSipMediaApplication_endpoints - The new set of endpoints for the specified SIP media application.

$sel:sipMediaApplicationId:UpdateSipMediaApplication', updateSipMediaApplication_sipMediaApplicationId - The SIP media application ID.

Request Lenses

updateSipMediaApplication_name :: Lens' UpdateSipMediaApplication (Maybe Text)

The new name for the specified SIP media application.

updateSipMediaApplication_endpoints :: Lens' UpdateSipMediaApplication (Maybe (NonEmpty SipMediaApplicationEndpoint))

The new set of endpoints for the specified SIP media application.

updateSipMediaApplication_sipMediaApplicationId :: Lens' UpdateSipMediaApplication Text

The SIP media application ID.

Destructuring the Response

data UpdateSipMediaApplicationResponse

See: newUpdateSipMediaApplicationResponse smart constructor.

Constructors

UpdateSipMediaApplicationResponse'

Fields

sipMediaApplication :: Maybe SipMediaApplication

The updated SIP media application details.

httpStatus :: Int

The response's http status code.

newUpdateSipMediaApplicationResponse

Arguments

:: Int
$sel:httpStatus:UpdateSipMediaApplicationResponse'

-> UpdateSipMediaApplicationResponse

Create a value of UpdateSipMediaApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sipMediaApplication:UpdateSipMediaApplicationResponse', updateSipMediaApplicationResponse_sipMediaApplication - The updated SIP media application details.

$sel:httpStatus:UpdateSipMediaApplicationResponse', updateSipMediaApplicationResponse_httpStatus - The response's http status code.

Response Lenses

updateSipMediaApplicationResponse_sipMediaApplication :: Lens' UpdateSipMediaApplicationResponse (Maybe SipMediaApplication)

The updated SIP media application details.

updateSipMediaApplicationResponse_httpStatus :: Lens' UpdateSipMediaApplicationResponse Int

The response's http status code.