Network.AWS.Chime.UpdateVoiceConnectorGroup

Description

Updates details of the specified Amazon Chime Voice Connector group, such as the name and Amazon Chime Voice Connector priority ranking.

Creating a Request

data UpdateVoiceConnectorGroup

See: newUpdateVoiceConnectorGroup smart constructor.

Constructors

UpdateVoiceConnectorGroup'

Fields

voiceConnectorGroupId :: Text

The Amazon Chime Voice Connector group ID.

name :: Text

The name of the Amazon Chime Voice Connector group.

voiceConnectorItems :: [VoiceConnectorItem]

The VoiceConnectorItems to associate with the group.

newUpdateVoiceConnectorGroup

Arguments

:: Text
$sel:voiceConnectorGroupId:UpdateVoiceConnectorGroup'

-> Text
$sel:name:UpdateVoiceConnectorGroup'

-> UpdateVoiceConnectorGroup

Create a value of UpdateVoiceConnectorGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:voiceConnectorGroupId:UpdateVoiceConnectorGroup', updateVoiceConnectorGroup_voiceConnectorGroupId - The Amazon Chime Voice Connector group ID.

$sel:name:UpdateVoiceConnectorGroup', updateVoiceConnectorGroup_name - The name of the Amazon Chime Voice Connector group.

$sel:voiceConnectorItems:UpdateVoiceConnectorGroup', updateVoiceConnectorGroup_voiceConnectorItems - The VoiceConnectorItems to associate with the group.

Request Lenses

updateVoiceConnectorGroup_voiceConnectorGroupId :: Lens' UpdateVoiceConnectorGroup Text

The Amazon Chime Voice Connector group ID.

updateVoiceConnectorGroup_name :: Lens' UpdateVoiceConnectorGroup Text

The name of the Amazon Chime Voice Connector group.

updateVoiceConnectorGroup_voiceConnectorItems :: Lens' UpdateVoiceConnectorGroup [VoiceConnectorItem]

The VoiceConnectorItems to associate with the group.

Destructuring the Response

data UpdateVoiceConnectorGroupResponse

See: newUpdateVoiceConnectorGroupResponse smart constructor.

Constructors

UpdateVoiceConnectorGroupResponse'

Fields

voiceConnectorGroup :: Maybe VoiceConnectorGroup

The updated Amazon Chime Voice Connector group details.

httpStatus :: Int

The response's http status code.

newUpdateVoiceConnectorGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateVoiceConnectorGroupResponse'

-> UpdateVoiceConnectorGroupResponse

Create a value of UpdateVoiceConnectorGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:voiceConnectorGroup:UpdateVoiceConnectorGroupResponse', updateVoiceConnectorGroupResponse_voiceConnectorGroup - The updated Amazon Chime Voice Connector group details.

$sel:httpStatus:UpdateVoiceConnectorGroupResponse', updateVoiceConnectorGroupResponse_httpStatus - The response's http status code.

Response Lenses

updateVoiceConnectorGroupResponse_voiceConnectorGroup :: Lens' UpdateVoiceConnectorGroupResponse (Maybe VoiceConnectorGroup)

The updated Amazon Chime Voice Connector group details.

updateVoiceConnectorGroupResponse_httpStatus :: Lens' UpdateVoiceConnectorGroupResponse Int

The response's http status code.