Network.AWS.ChimeSDKIdentity

Description

Derived from API version 2021-04-20 of the AWS service descriptions, licensed under Apache 2.0.

The Amazon Chime SDK Identity APIs in this section allow software developers to create and manage unique instances of their messaging applications. These APIs provide the overarching framework for creating and sending messages. For more information about the identity APIs, refer to Amazon Chime SDK identity.

Service Configuration

defaultService :: Service

API version 2021-04-20 of the Amazon Chime SDK Identity SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ChimeSDKIdentity.

ThrottledClientException

_ThrottledClientException :: AsError a => Getting (First ServiceError) a ServiceError

The client exceeded its request rate limit.

ResourceLimitExceededException

_ResourceLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request exceeds the resource limit.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request could not be processed because of conflict in the current state of the resource.

ForbiddenException

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

The client is permanently forbidden from making the request.

ServiceFailureException

_ServiceFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The service encountered an unexpected error.

UnauthorizedClientException

_UnauthorizedClientException :: AsError a => Getting (First ServiceError) a ServiceError

The client is not currently authorized to make the request.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is currently unavailable.

BadRequestException

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The input parameters don't match the service's restrictions.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

CreateAppInstance

data CreateAppInstance

See: newCreateAppInstance smart constructor.

Constructors

CreateAppInstance' (Maybe (Sensitive Text)) (Maybe (NonEmpty Tag)) (Sensitive Text) (Sensitive Text)

newCreateAppInstance

Arguments

:: Text
$sel:name:CreateAppInstance'

-> Text
$sel:clientRequestToken:CreateAppInstance'

-> CreateAppInstance

Create a value of CreateAppInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadata:CreateAppInstance', createAppInstance_metadata - The metadata of the AppInstance. Limited to a 1KB string in UTF-8.

$sel:tags:CreateAppInstance', createAppInstance_tags - Tags assigned to the AppInstanceUser.

$sel:name:CreateAppInstance', createAppInstance_name - The name of the AppInstance.

$sel:clientRequestToken:CreateAppInstance', createAppInstance_clientRequestToken - The ClientRequestToken of the AppInstance.

data CreateAppInstanceResponse

See: newCreateAppInstanceResponse smart constructor.

Constructors

CreateAppInstanceResponse' (Maybe Text) Int

newCreateAppInstanceResponse

Arguments

:: Int
$sel:httpStatus:CreateAppInstanceResponse'

-> CreateAppInstanceResponse

Create a value of CreateAppInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceArn:CreateAppInstanceResponse', createAppInstanceResponse_appInstanceArn - The Amazon Resource Number (ARN) of the AppInstance.

$sel:httpStatus:CreateAppInstanceResponse', createAppInstanceResponse_httpStatus - The response's http status code.

GetAppInstanceRetentionSettings

data GetAppInstanceRetentionSettings

See: newGetAppInstanceRetentionSettings smart constructor.

Constructors

GetAppInstanceRetentionSettings' Text

newGetAppInstanceRetentionSettings

Arguments

:: Text
$sel:appInstanceArn:GetAppInstanceRetentionSettings'

-> GetAppInstanceRetentionSettings

Create a value of GetAppInstanceRetentionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceArn:GetAppInstanceRetentionSettings', getAppInstanceRetentionSettings_appInstanceArn - The ARN of the AppInstance.

data GetAppInstanceRetentionSettingsResponse

See: newGetAppInstanceRetentionSettingsResponse smart constructor.

Constructors

GetAppInstanceRetentionSettingsResponse' (Maybe AppInstanceRetentionSettings) (Maybe POSIX) Int

newGetAppInstanceRetentionSettingsResponse

Arguments

:: Int
$sel:httpStatus:GetAppInstanceRetentionSettingsResponse'

-> GetAppInstanceRetentionSettingsResponse

Create a value of GetAppInstanceRetentionSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceRetentionSettings:GetAppInstanceRetentionSettingsResponse', getAppInstanceRetentionSettingsResponse_appInstanceRetentionSettings - The retention settings for the AppInstance.

$sel:initiateDeletionTimestamp:GetAppInstanceRetentionSettingsResponse', getAppInstanceRetentionSettingsResponse_initiateDeletionTimestamp - The timestamp representing the time at which the specified items are retained, in Epoch Seconds.

$sel:httpStatus:GetAppInstanceRetentionSettingsResponse', getAppInstanceRetentionSettingsResponse_httpStatus - The response's http status code.

DescribeAppInstanceAdmin

data DescribeAppInstanceAdmin

See: newDescribeAppInstanceAdmin smart constructor.

Constructors

DescribeAppInstanceAdmin' Text Text

newDescribeAppInstanceAdmin

Arguments

:: Text
$sel:appInstanceAdminArn:DescribeAppInstanceAdmin'

-> Text
$sel:appInstanceArn:DescribeAppInstanceAdmin'

-> DescribeAppInstanceAdmin

Create a value of DescribeAppInstanceAdmin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceAdminArn:DescribeAppInstanceAdmin', describeAppInstanceAdmin_appInstanceAdminArn - The ARN of the AppInstanceAdmin.

$sel:appInstanceArn:DescribeAppInstanceAdmin', describeAppInstanceAdmin_appInstanceArn - The ARN of the AppInstance.

data DescribeAppInstanceAdminResponse

See: newDescribeAppInstanceAdminResponse smart constructor.

Constructors

DescribeAppInstanceAdminResponse' (Maybe AppInstanceAdmin) Int

newDescribeAppInstanceAdminResponse

Arguments

:: Int
$sel:httpStatus:DescribeAppInstanceAdminResponse'

-> DescribeAppInstanceAdminResponse

Create a value of DescribeAppInstanceAdminResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceAdmin:DescribeAppInstanceAdminResponse', describeAppInstanceAdminResponse_appInstanceAdmin - The ARN and name of the AppInstanceUser, the ARN of the AppInstance, and the created and last-updated timestamps. All timestamps use epoch milliseconds.

$sel:httpStatus:DescribeAppInstanceAdminResponse', describeAppInstanceAdminResponse_httpStatus - The response's http status code.

CreateAppInstanceUser

data CreateAppInstanceUser

See: newCreateAppInstanceUser smart constructor.

Constructors

CreateAppInstanceUser' (Maybe (Sensitive Text)) (Maybe (NonEmpty Tag)) Text (Sensitive Text) (Sensitive Text) (Sensitive Text)

newCreateAppInstanceUser

Arguments

:: Text
$sel:appInstanceArn:CreateAppInstanceUser'

-> Text
$sel:appInstanceUserId:CreateAppInstanceUser'

-> Text
$sel:name:CreateAppInstanceUser'

-> Text
$sel:clientRequestToken:CreateAppInstanceUser'

-> CreateAppInstanceUser

Create a value of CreateAppInstanceUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadata:CreateAppInstanceUser', createAppInstanceUser_metadata - The request's metadata. Limited to a 1KB string in UTF-8.

$sel:tags:CreateAppInstanceUser', createAppInstanceUser_tags - Tags assigned to the AppInstanceUser.

$sel:appInstanceArn:CreateAppInstanceUser', createAppInstanceUser_appInstanceArn - The ARN of the AppInstance request.

$sel:appInstanceUserId:CreateAppInstanceUser', createAppInstanceUser_appInstanceUserId - The user ID of the AppInstance.

$sel:name:CreateAppInstanceUser', createAppInstanceUser_name - The user's name.

$sel:clientRequestToken:CreateAppInstanceUser', createAppInstanceUser_clientRequestToken - The token assigned to the user requesting an AppInstance.

data CreateAppInstanceUserResponse

See: newCreateAppInstanceUserResponse smart constructor.

Constructors

CreateAppInstanceUserResponse' (Maybe Text) Int

newCreateAppInstanceUserResponse

Arguments

:: Int
$sel:httpStatus:CreateAppInstanceUserResponse'

-> CreateAppInstanceUserResponse

Create a value of CreateAppInstanceUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUserArn:CreateAppInstanceUserResponse', createAppInstanceUserResponse_appInstanceUserArn - The user's ARN.

$sel:httpStatus:CreateAppInstanceUserResponse', createAppInstanceUserResponse_httpStatus - The response's http status code.

PutAppInstanceRetentionSettings

data PutAppInstanceRetentionSettings

See: newPutAppInstanceRetentionSettings smart constructor.

Constructors

PutAppInstanceRetentionSettings' Text AppInstanceRetentionSettings

newPutAppInstanceRetentionSettings

Arguments

:: Text
$sel:appInstanceArn:PutAppInstanceRetentionSettings'

-> AppInstanceRetentionSettings
$sel:appInstanceRetentionSettings:PutAppInstanceRetentionSettings'

-> PutAppInstanceRetentionSettings

Create a value of PutAppInstanceRetentionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceArn:PutAppInstanceRetentionSettings', putAppInstanceRetentionSettings_appInstanceArn - The ARN of the AppInstance.

$sel:appInstanceRetentionSettings:PutAppInstanceRetentionSettings', putAppInstanceRetentionSettings_appInstanceRetentionSettings - The time in days to retain data. Data type: number.

data PutAppInstanceRetentionSettingsResponse

See: newPutAppInstanceRetentionSettingsResponse smart constructor.

Constructors

PutAppInstanceRetentionSettingsResponse' (Maybe AppInstanceRetentionSettings) (Maybe POSIX) Int

newPutAppInstanceRetentionSettingsResponse

Arguments

:: Int
$sel:httpStatus:PutAppInstanceRetentionSettingsResponse'

-> PutAppInstanceRetentionSettingsResponse

Create a value of PutAppInstanceRetentionSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceRetentionSettings:PutAppInstanceRetentionSettings', putAppInstanceRetentionSettingsResponse_appInstanceRetentionSettings - The time in days to retain data. Data type: number.

$sel:initiateDeletionTimestamp:PutAppInstanceRetentionSettingsResponse', putAppInstanceRetentionSettingsResponse_initiateDeletionTimestamp - The time at which the API deletes data.

$sel:httpStatus:PutAppInstanceRetentionSettingsResponse', putAppInstanceRetentionSettingsResponse_httpStatus - The response's http status code.

CreateAppInstanceAdmin

data CreateAppInstanceAdmin

See: newCreateAppInstanceAdmin smart constructor.

Constructors

CreateAppInstanceAdmin' Text Text

newCreateAppInstanceAdmin

Arguments

:: Text
$sel:appInstanceAdminArn:CreateAppInstanceAdmin'

-> Text
$sel:appInstanceArn:CreateAppInstanceAdmin'

-> CreateAppInstanceAdmin

Create a value of CreateAppInstanceAdmin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceAdminArn:CreateAppInstanceAdmin', createAppInstanceAdmin_appInstanceAdminArn - The ARN of the administrator of the current AppInstance.

$sel:appInstanceArn:CreateAppInstanceAdmin', createAppInstanceAdmin_appInstanceArn - The ARN of the AppInstance.

data CreateAppInstanceAdminResponse

See: newCreateAppInstanceAdminResponse smart constructor.

Constructors

CreateAppInstanceAdminResponse' (Maybe Identity) (Maybe Text) Int

newCreateAppInstanceAdminResponse

Arguments

:: Int
$sel:httpStatus:CreateAppInstanceAdminResponse'

-> CreateAppInstanceAdminResponse

Create a value of CreateAppInstanceAdminResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceAdmin:CreateAppInstanceAdminResponse', createAppInstanceAdminResponse_appInstanceAdmin - The name and ARN of the admin for the AppInstance.

$sel:appInstanceArn:CreateAppInstanceAdmin', createAppInstanceAdminResponse_appInstanceArn - The ARN of the of the admin for the AppInstance.

$sel:httpStatus:CreateAppInstanceAdminResponse', createAppInstanceAdminResponse_httpStatus - The response's http status code.

ListAppInstanceAdmins

data ListAppInstanceAdmins

See: newListAppInstanceAdmins smart constructor.

Constructors

ListAppInstanceAdmins' (Maybe (Sensitive Text)) (Maybe Natural) Text

newListAppInstanceAdmins

Arguments

:: Text
$sel:appInstanceArn:ListAppInstanceAdmins'

-> ListAppInstanceAdmins

Create a value of ListAppInstanceAdmins with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppInstanceAdmins', listAppInstanceAdmins_nextToken - The token returned from previous API requests until the number of administrators is reached.

$sel:maxResults:ListAppInstanceAdmins', listAppInstanceAdmins_maxResults - The maximum number of administrators that you want to return.

$sel:appInstanceArn:ListAppInstanceAdmins', listAppInstanceAdmins_appInstanceArn - The ARN of the AppInstance.

data ListAppInstanceAdminsResponse

See: newListAppInstanceAdminsResponse smart constructor.

Constructors

ListAppInstanceAdminsResponse' (Maybe (Sensitive Text)) (Maybe [AppInstanceAdminSummary]) (Maybe Text) Int

newListAppInstanceAdminsResponse

Arguments

:: Int
$sel:httpStatus:ListAppInstanceAdminsResponse'

-> ListAppInstanceAdminsResponse

Create a value of ListAppInstanceAdminsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppInstanceAdmins', listAppInstanceAdminsResponse_nextToken - The token returned from previous API requests until the number of administrators is reached.

$sel:appInstanceAdmins:ListAppInstanceAdminsResponse', listAppInstanceAdminsResponse_appInstanceAdmins - The information for each administrator.

$sel:appInstanceArn:ListAppInstanceAdmins', listAppInstanceAdminsResponse_appInstanceArn - The ARN of the AppInstance.

$sel:httpStatus:ListAppInstanceAdminsResponse', listAppInstanceAdminsResponse_httpStatus - The response's http status code.

ListAppInstances

data ListAppInstances

See: newListAppInstances smart constructor.

Constructors

ListAppInstances' (Maybe (Sensitive Text)) (Maybe Natural)

newListAppInstances :: ListAppInstances

Create a value of ListAppInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppInstances', listAppInstances_nextToken - The token passed by previous API requests until you reach the maximum number of AppInstances.

$sel:maxResults:ListAppInstances', listAppInstances_maxResults - The maximum number of AppInstances that you want to return.

data ListAppInstancesResponse

See: newListAppInstancesResponse smart constructor.

Constructors

ListAppInstancesResponse' (Maybe [AppInstanceSummary]) (Maybe (Sensitive Text)) Int

newListAppInstancesResponse

Arguments

:: Int
$sel:httpStatus:ListAppInstancesResponse'

-> ListAppInstancesResponse

Create a value of ListAppInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstances:ListAppInstancesResponse', listAppInstancesResponse_appInstances - The information for each AppInstance.

$sel:nextToken:ListAppInstances', listAppInstancesResponse_nextToken - The token passed by previous API requests until the maximum number of AppInstances is reached.

$sel:httpStatus:ListAppInstancesResponse', listAppInstancesResponse_httpStatus - The response's http status code.

DescribeAppInstanceUser

data DescribeAppInstanceUser

See: newDescribeAppInstanceUser smart constructor.

Constructors

DescribeAppInstanceUser' Text

newDescribeAppInstanceUser

Arguments

:: Text
$sel:appInstanceUserArn:DescribeAppInstanceUser'

-> DescribeAppInstanceUser

Create a value of DescribeAppInstanceUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUserArn:DescribeAppInstanceUser', describeAppInstanceUser_appInstanceUserArn - The ARN of the AppInstanceUser.

data DescribeAppInstanceUserResponse

See: newDescribeAppInstanceUserResponse smart constructor.

Constructors

DescribeAppInstanceUserResponse' (Maybe AppInstanceUser) Int

newDescribeAppInstanceUserResponse

Arguments

:: Int
$sel:httpStatus:DescribeAppInstanceUserResponse'

-> DescribeAppInstanceUserResponse

Create a value of DescribeAppInstanceUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUser:DescribeAppInstanceUserResponse', describeAppInstanceUserResponse_appInstanceUser - The name of the AppInstanceUser.

$sel:httpStatus:DescribeAppInstanceUserResponse', describeAppInstanceUserResponse_httpStatus - The response's http status code.

DescribeAppInstance

data DescribeAppInstance

See: newDescribeAppInstance smart constructor.

Constructors

DescribeAppInstance' Text

newDescribeAppInstance

Arguments

:: Text
$sel:appInstanceArn:DescribeAppInstance'

-> DescribeAppInstance

Create a value of DescribeAppInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceArn:DescribeAppInstance', describeAppInstance_appInstanceArn - The ARN of the AppInstance.

data DescribeAppInstanceResponse

See: newDescribeAppInstanceResponse smart constructor.

Constructors

DescribeAppInstanceResponse' (Maybe AppInstance) Int

newDescribeAppInstanceResponse

Arguments

:: Int
$sel:httpStatus:DescribeAppInstanceResponse'

-> DescribeAppInstanceResponse

Create a value of DescribeAppInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstance:DescribeAppInstanceResponse', describeAppInstanceResponse_appInstance - The ARN, metadata, created and last-updated timestamps, and the name of the AppInstance. All timestamps use epoch milliseconds.

$sel:httpStatus:DescribeAppInstanceResponse', describeAppInstanceResponse_httpStatus - The response's http status code.

ListAppInstanceUsers

data ListAppInstanceUsers

See: newListAppInstanceUsers smart constructor.

Constructors

ListAppInstanceUsers' (Maybe (Sensitive Text)) (Maybe Natural) Text

newListAppInstanceUsers

Arguments

:: Text
$sel:appInstanceArn:ListAppInstanceUsers'

-> ListAppInstanceUsers

Create a value of ListAppInstanceUsers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppInstanceUsers', listAppInstanceUsers_nextToken - The token passed by previous API calls until all requested users are returned.

$sel:maxResults:ListAppInstanceUsers', listAppInstanceUsers_maxResults - The maximum number of requests that you want returned.

$sel:appInstanceArn:ListAppInstanceUsers', listAppInstanceUsers_appInstanceArn - The ARN of the AppInstance.

data ListAppInstanceUsersResponse

See: newListAppInstanceUsersResponse smart constructor.

Constructors

ListAppInstanceUsersResponse' (Maybe (Sensitive Text)) (Maybe [AppInstanceUserSummary]) (Maybe Text) Int

newListAppInstanceUsersResponse

Arguments

:: Int
$sel:httpStatus:ListAppInstanceUsersResponse'

-> ListAppInstanceUsersResponse

Create a value of ListAppInstanceUsersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppInstanceUsers', listAppInstanceUsersResponse_nextToken - The token passed by previous API calls until all requested users are returned.

$sel:appInstanceUsers:ListAppInstanceUsersResponse', listAppInstanceUsersResponse_appInstanceUsers - The information for each requested AppInstanceUser.

$sel:appInstanceArn:ListAppInstanceUsers', listAppInstanceUsersResponse_appInstanceArn - The ARN of the AppInstance.

$sel:httpStatus:ListAppInstanceUsersResponse', listAppInstanceUsersResponse_httpStatus - The response's http status code.

DeleteAppInstanceUser

data DeleteAppInstanceUser

See: newDeleteAppInstanceUser smart constructor.

Constructors

DeleteAppInstanceUser' Text

newDeleteAppInstanceUser

Arguments

:: Text
$sel:appInstanceUserArn:DeleteAppInstanceUser'

-> DeleteAppInstanceUser

Create a value of DeleteAppInstanceUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUserArn:DeleteAppInstanceUser', deleteAppInstanceUser_appInstanceUserArn - The ARN of the user request being deleted.

data DeleteAppInstanceUserResponse

See: newDeleteAppInstanceUserResponse smart constructor.

Constructors

DeleteAppInstanceUserResponse'

newDeleteAppInstanceUserResponse :: DeleteAppInstanceUserResponse

Create a value of DeleteAppInstanceUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

UpdateAppInstanceUser

data UpdateAppInstanceUser

See: newUpdateAppInstanceUser smart constructor.

Constructors

UpdateAppInstanceUser' Text (Sensitive Text) (Sensitive Text)

newUpdateAppInstanceUser

Arguments

:: Text
$sel:appInstanceUserArn:UpdateAppInstanceUser'

-> Text
$sel:name:UpdateAppInstanceUser'

-> Text
$sel:metadata:UpdateAppInstanceUser'

-> UpdateAppInstanceUser

Create a value of UpdateAppInstanceUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUserArn:UpdateAppInstanceUser', updateAppInstanceUser_appInstanceUserArn - The ARN of the AppInstanceUser.

$sel:name:UpdateAppInstanceUser', updateAppInstanceUser_name - The name of the AppInstanceUser.

$sel:metadata:UpdateAppInstanceUser', updateAppInstanceUser_metadata - The metadata of the AppInstanceUser.

data UpdateAppInstanceUserResponse

See: newUpdateAppInstanceUserResponse smart constructor.

Constructors

UpdateAppInstanceUserResponse' (Maybe Text) Int

newUpdateAppInstanceUserResponse

Arguments

:: Int
$sel:httpStatus:UpdateAppInstanceUserResponse'

-> UpdateAppInstanceUserResponse

Create a value of UpdateAppInstanceUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUserArn:UpdateAppInstanceUser', updateAppInstanceUserResponse_appInstanceUserArn - The ARN of the AppInstanceUser.

$sel:httpStatus:UpdateAppInstanceUserResponse', updateAppInstanceUserResponse_httpStatus - The response's http status code.

DeleteAppInstanceAdmin

data DeleteAppInstanceAdmin

See: newDeleteAppInstanceAdmin smart constructor.

Constructors

DeleteAppInstanceAdmin' Text Text

newDeleteAppInstanceAdmin

Arguments

:: Text
$sel:appInstanceAdminArn:DeleteAppInstanceAdmin'

-> Text
$sel:appInstanceArn:DeleteAppInstanceAdmin'

-> DeleteAppInstanceAdmin

Create a value of DeleteAppInstanceAdmin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceAdminArn:DeleteAppInstanceAdmin', deleteAppInstanceAdmin_appInstanceAdminArn - The ARN of the AppInstance's administrator.

$sel:appInstanceArn:DeleteAppInstanceAdmin', deleteAppInstanceAdmin_appInstanceArn - The ARN of the AppInstance.

data DeleteAppInstanceAdminResponse

See: newDeleteAppInstanceAdminResponse smart constructor.

Constructors

DeleteAppInstanceAdminResponse'

newDeleteAppInstanceAdminResponse :: DeleteAppInstanceAdminResponse

Create a value of DeleteAppInstanceAdminResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

DeleteAppInstance

data DeleteAppInstance

See: newDeleteAppInstance smart constructor.

Constructors

DeleteAppInstance' Text

newDeleteAppInstance

Arguments

:: Text
$sel:appInstanceArn:DeleteAppInstance'

-> DeleteAppInstance

Create a value of DeleteAppInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceArn:DeleteAppInstance', deleteAppInstance_appInstanceArn - The ARN of the AppInstance.

data DeleteAppInstanceResponse

See: newDeleteAppInstanceResponse smart constructor.

Constructors

DeleteAppInstanceResponse'

newDeleteAppInstanceResponse :: DeleteAppInstanceResponse

Create a value of DeleteAppInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

UpdateAppInstance

data UpdateAppInstance

See: newUpdateAppInstance smart constructor.

Constructors

UpdateAppInstance' Text (Sensitive Text) (Sensitive Text)

newUpdateAppInstance

Arguments

:: Text
$sel:appInstanceArn:UpdateAppInstance'

-> Text
$sel:name:UpdateAppInstance'

-> Text
$sel:metadata:UpdateAppInstance'

-> UpdateAppInstance

Create a value of UpdateAppInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceArn:UpdateAppInstance', updateAppInstance_appInstanceArn - The ARN of the AppInstance.

$sel:name:UpdateAppInstance', updateAppInstance_name - The name that you want to change.

$sel:metadata:UpdateAppInstance', updateAppInstance_metadata - The metadata that you want to change.

data UpdateAppInstanceResponse

See: newUpdateAppInstanceResponse smart constructor.

Constructors

UpdateAppInstanceResponse' (Maybe Text) Int

newUpdateAppInstanceResponse

Arguments

:: Int
$sel:httpStatus:UpdateAppInstanceResponse'

-> UpdateAppInstanceResponse

Create a value of UpdateAppInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceArn:UpdateAppInstance', updateAppInstanceResponse_appInstanceArn - The ARN of the AppInstance.

$sel:httpStatus:UpdateAppInstanceResponse', updateAppInstanceResponse_httpStatus - The response's http status code.

Types

AppInstance

data AppInstance

The details of an AppInstance, an instance of an Amazon Chime SDK messaging application.

See: newAppInstance smart constructor.

Constructors

AppInstance' (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) (Maybe Text) (Maybe POSIX) (Maybe POSIX)

newAppInstance :: AppInstance

Create a value of AppInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:AppInstance', appInstance_name - The name of an AppInstance.

$sel:metadata:AppInstance', appInstance_metadata - The metadata of an AppInstance.

$sel:appInstanceArn:AppInstance', appInstance_appInstanceArn - The ARN of the messaging instance.

$sel:createdTimestamp:AppInstance', appInstance_createdTimestamp - The time at which an AppInstance was created. In epoch milliseconds.

$sel:lastUpdatedTimestamp:AppInstance', appInstance_lastUpdatedTimestamp - The time an AppInstance was last updated. In epoch milliseconds.

AppInstanceAdmin

data AppInstanceAdmin

The details of an AppInstanceAdmin.

See: newAppInstanceAdmin smart constructor.

Constructors

AppInstanceAdmin' (Maybe Identity) (Maybe Text) (Maybe POSIX)

newAppInstanceAdmin :: AppInstanceAdmin

Create a value of AppInstanceAdmin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:admin:AppInstanceAdmin', appInstanceAdmin_admin - The AppInstanceAdmin data.

$sel:appInstanceArn:AppInstanceAdmin', appInstanceAdmin_appInstanceArn - The ARN of the AppInstance for which the user is an administrator.

$sel:createdTimestamp:AppInstanceAdmin', appInstanceAdmin_createdTimestamp - The time at which an administrator was created.

AppInstanceAdminSummary

data AppInstanceAdminSummary

Summary of the details of an AppInstanceAdmin.

See: newAppInstanceAdminSummary smart constructor.

Constructors

AppInstanceAdminSummary' (Maybe Identity)

newAppInstanceAdminSummary :: AppInstanceAdminSummary

Create a value of AppInstanceAdminSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:admin:AppInstanceAdminSummary', appInstanceAdminSummary_admin - The details of the AppInstanceAdmin.

AppInstanceRetentionSettings

data AppInstanceRetentionSettings

The details of the data-retention settings for an AppInstance.

See: newAppInstanceRetentionSettings smart constructor.

Constructors

AppInstanceRetentionSettings' (Maybe ChannelRetentionSettings)

newAppInstanceRetentionSettings :: AppInstanceRetentionSettings

Create a value of AppInstanceRetentionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelRetentionSettings:AppInstanceRetentionSettings', appInstanceRetentionSettings_channelRetentionSettings - The length of time in days to retain the messages in a channel.

AppInstanceSummary

data AppInstanceSummary

Summary of the data for an AppInstance.

See: newAppInstanceSummary smart constructor.

Constructors

AppInstanceSummary' (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) (Maybe Text)

newAppInstanceSummary :: AppInstanceSummary

Create a value of AppInstanceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:AppInstanceSummary', appInstanceSummary_name - The name of the AppInstance.

$sel:metadata:AppInstanceSummary', appInstanceSummary_metadata - The metadata of the AppInstance.

$sel:appInstanceArn:AppInstanceSummary', appInstanceSummary_appInstanceArn - The AppInstance ARN.

AppInstanceUser

data AppInstanceUser

The details of an AppInstanceUser.

See: newAppInstanceUser smart constructor.

Constructors

AppInstanceUser' (Maybe Text) (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) (Maybe POSIX) (Maybe POSIX)

newAppInstanceUser :: AppInstanceUser

Create a value of AppInstanceUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUserArn:AppInstanceUser', appInstanceUser_appInstanceUserArn - The ARN of the AppInstanceUser.

$sel:name:AppInstanceUser', appInstanceUser_name - The name of the AppInstanceUser.

$sel:metadata:AppInstanceUser', appInstanceUser_metadata - The metadata of the AppInstanceUser.

$sel:createdTimestamp:AppInstanceUser', appInstanceUser_createdTimestamp - The time at which the AppInstanceUser was created.

$sel:lastUpdatedTimestamp:AppInstanceUser', appInstanceUser_lastUpdatedTimestamp - The time at which the AppInstanceUser was last updated.

AppInstanceUserSummary

data AppInstanceUserSummary

Summary of the details of an AppInstanceUser.

See: newAppInstanceUserSummary smart constructor.

Constructors

AppInstanceUserSummary' (Maybe Text) (Maybe (Sensitive Text)) (Maybe (Sensitive Text))

newAppInstanceUserSummary :: AppInstanceUserSummary

Create a value of AppInstanceUserSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUserArn:AppInstanceUserSummary', appInstanceUserSummary_appInstanceUserArn - The ARN of the AppInstanceUser.

$sel:name:AppInstanceUserSummary', appInstanceUserSummary_name - The name of an AppInstanceUser.

$sel:metadata:AppInstanceUserSummary', appInstanceUserSummary_metadata - The metadata of the AppInstanceUser.

ChannelRetentionSettings

data ChannelRetentionSettings

The details of the retention settings for a channel.

See: newChannelRetentionSettings smart constructor.

Constructors

ChannelRetentionSettings' (Maybe Natural)

newChannelRetentionSettings :: ChannelRetentionSettings

Create a value of ChannelRetentionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retentionDays:ChannelRetentionSettings', channelRetentionSettings_retentionDays - The time in days to retain the messages in a channel.

Identity

data Identity

The details of a user.

See: newIdentity smart constructor.

Constructors

Identity' (Maybe Text) (Maybe (Sensitive Text))

newIdentity :: Identity

Create a value of Identity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Identity', identity_arn - The ARN in an Identity.

$sel:name:Identity', identity_name - The name in an Identity.

Tag

data Tag

Describes a tag applied to a resource.

See: newTag smart constructor.

Constructors

Tag' (Sensitive Text) (Sensitive Text)

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key of the tag.

$sel:value:Tag', tag_value - The value of the tag.