Network.AWS.ChimeSDKIdentity.ListAppInstanceAdmins

Description

Returns a list of the administrators in the AppInstance.

Creating a Request

data ListAppInstanceAdmins

See: newListAppInstanceAdmins smart constructor.

Constructors

ListAppInstanceAdmins'

Fields

nextToken :: Maybe (Sensitive Text)

The token returned from previous API requests until the number of administrators is reached.

maxResults :: Maybe Natural

The maximum number of administrators that you want to return.

appInstanceArn :: Text

The ARN of the AppInstance.

newListAppInstanceAdmins

Arguments

:: Text
$sel:appInstanceArn:ListAppInstanceAdmins'

-> ListAppInstanceAdmins

Create a value of ListAppInstanceAdmins with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppInstanceAdmins', listAppInstanceAdmins_nextToken - The token returned from previous API requests until the number of administrators is reached.

$sel:maxResults:ListAppInstanceAdmins', listAppInstanceAdmins_maxResults - The maximum number of administrators that you want to return.

$sel:appInstanceArn:ListAppInstanceAdmins', listAppInstanceAdmins_appInstanceArn - The ARN of the AppInstance.

Request Lenses

listAppInstanceAdmins_nextToken :: Lens' ListAppInstanceAdmins (Maybe Text)

The token returned from previous API requests until the number of administrators is reached.

listAppInstanceAdmins_maxResults :: Lens' ListAppInstanceAdmins (Maybe Natural)

The maximum number of administrators that you want to return.

listAppInstanceAdmins_appInstanceArn :: Lens' ListAppInstanceAdmins Text

The ARN of the AppInstance.

Destructuring the Response

data ListAppInstanceAdminsResponse

See: newListAppInstanceAdminsResponse smart constructor.

Constructors

ListAppInstanceAdminsResponse'

Fields

nextToken :: Maybe (Sensitive Text)

The token returned from previous API requests until the number of administrators is reached.

appInstanceAdmins :: Maybe [AppInstanceAdminSummary]

The information for each administrator.

appInstanceArn :: Maybe Text

The ARN of the AppInstance.

httpStatus :: Int

The response's http status code.

newListAppInstanceAdminsResponse

Arguments

:: Int
$sel:httpStatus:ListAppInstanceAdminsResponse'

-> ListAppInstanceAdminsResponse

Create a value of ListAppInstanceAdminsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAppInstanceAdmins', listAppInstanceAdminsResponse_nextToken - The token returned from previous API requests until the number of administrators is reached.

$sel:appInstanceAdmins:ListAppInstanceAdminsResponse', listAppInstanceAdminsResponse_appInstanceAdmins - The information for each administrator.

$sel:appInstanceArn:ListAppInstanceAdmins', listAppInstanceAdminsResponse_appInstanceArn - The ARN of the AppInstance.

$sel:httpStatus:ListAppInstanceAdminsResponse', listAppInstanceAdminsResponse_httpStatus - The response's http status code.

Response Lenses

listAppInstanceAdminsResponse_nextToken :: Lens' ListAppInstanceAdminsResponse (Maybe Text)

The token returned from previous API requests until the number of administrators is reached.

listAppInstanceAdminsResponse_appInstanceAdmins :: Lens' ListAppInstanceAdminsResponse (Maybe [AppInstanceAdminSummary])

The information for each administrator.

listAppInstanceAdminsResponse_appInstanceArn :: Lens' ListAppInstanceAdminsResponse (Maybe Text)

The ARN of the AppInstance.

listAppInstanceAdminsResponse_httpStatus :: Lens' ListAppInstanceAdminsResponse Int

The response's http status code.