Network.AWS.ChimeSDKIdentity.Types

Description

Service Configuration

defaultService :: Service

API version 2021-04-20 of the Amazon Chime SDK Identity SDK configuration.

Errors

_ThrottledClientException :: AsError a => Getting (First ServiceError) a ServiceError

The client exceeded its request rate limit.

_ResourceLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request exceeds the resource limit.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request could not be processed because of conflict in the current state of the resource.

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

The client is permanently forbidden from making the request.

_ServiceFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The service encountered an unexpected error.

_UnauthorizedClientException :: AsError a => Getting (First ServiceError) a ServiceError

The client is not currently authorized to make the request.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is currently unavailable.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The input parameters don't match the service's restrictions.

AppInstance

data AppInstance

The details of an AppInstance, an instance of an Amazon Chime SDK messaging application.

See: newAppInstance smart constructor.

Constructors

AppInstance'

Fields

name :: Maybe (Sensitive Text)

The name of an AppInstance.

metadata :: Maybe (Sensitive Text)

The metadata of an AppInstance.

appInstanceArn :: Maybe Text

The ARN of the messaging instance.

createdTimestamp :: Maybe POSIX

The time at which an AppInstance was created. In epoch milliseconds.

lastUpdatedTimestamp :: Maybe POSIX

The time an AppInstance was last updated. In epoch milliseconds.

newAppInstance :: AppInstance

Create a value of AppInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:AppInstance', appInstance_name - The name of an AppInstance.

$sel:metadata:AppInstance', appInstance_metadata - The metadata of an AppInstance.

$sel:appInstanceArn:AppInstance', appInstance_appInstanceArn - The ARN of the messaging instance.

$sel:createdTimestamp:AppInstance', appInstance_createdTimestamp - The time at which an AppInstance was created. In epoch milliseconds.

$sel:lastUpdatedTimestamp:AppInstance', appInstance_lastUpdatedTimestamp - The time an AppInstance was last updated. In epoch milliseconds.

appInstance_name :: Lens' AppInstance (Maybe Text)

The name of an AppInstance.

appInstance_metadata :: Lens' AppInstance (Maybe Text)

The metadata of an AppInstance.

appInstance_appInstanceArn :: Lens' AppInstance (Maybe Text)

The ARN of the messaging instance.

appInstance_createdTimestamp :: Lens' AppInstance (Maybe UTCTime)

The time at which an AppInstance was created. In epoch milliseconds.

appInstance_lastUpdatedTimestamp :: Lens' AppInstance (Maybe UTCTime)

The time an AppInstance was last updated. In epoch milliseconds.

AppInstanceAdmin

data AppInstanceAdmin

The details of an AppInstanceAdmin.

See: newAppInstanceAdmin smart constructor.

Constructors

AppInstanceAdmin'

Fields

admin :: Maybe Identity

The AppInstanceAdmin data.

appInstanceArn :: Maybe Text

The ARN of the AppInstance for which the user is an administrator.

createdTimestamp :: Maybe POSIX

The time at which an administrator was created.

newAppInstanceAdmin :: AppInstanceAdmin

Create a value of AppInstanceAdmin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:admin:AppInstanceAdmin', appInstanceAdmin_admin - The AppInstanceAdmin data.

$sel:appInstanceArn:AppInstanceAdmin', appInstanceAdmin_appInstanceArn - The ARN of the AppInstance for which the user is an administrator.

$sel:createdTimestamp:AppInstanceAdmin', appInstanceAdmin_createdTimestamp - The time at which an administrator was created.

appInstanceAdmin_admin :: Lens' AppInstanceAdmin (Maybe Identity)

The AppInstanceAdmin data.

appInstanceAdmin_appInstanceArn :: Lens' AppInstanceAdmin (Maybe Text)

The ARN of the AppInstance for which the user is an administrator.

appInstanceAdmin_createdTimestamp :: Lens' AppInstanceAdmin (Maybe UTCTime)

The time at which an administrator was created.

AppInstanceAdminSummary

data AppInstanceAdminSummary

Summary of the details of an AppInstanceAdmin.

See: newAppInstanceAdminSummary smart constructor.

Constructors

AppInstanceAdminSummary'

Fields

admin :: Maybe Identity

The details of the AppInstanceAdmin.

newAppInstanceAdminSummary :: AppInstanceAdminSummary

Create a value of AppInstanceAdminSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:admin:AppInstanceAdminSummary', appInstanceAdminSummary_admin - The details of the AppInstanceAdmin.

appInstanceAdminSummary_admin :: Lens' AppInstanceAdminSummary (Maybe Identity)

The details of the AppInstanceAdmin.

AppInstanceRetentionSettings

data AppInstanceRetentionSettings

The details of the data-retention settings for an AppInstance.

See: newAppInstanceRetentionSettings smart constructor.

Constructors

AppInstanceRetentionSettings'

Fields

channelRetentionSettings :: Maybe ChannelRetentionSettings

The length of time in days to retain the messages in a channel.

newAppInstanceRetentionSettings :: AppInstanceRetentionSettings

Create a value of AppInstanceRetentionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelRetentionSettings:AppInstanceRetentionSettings', appInstanceRetentionSettings_channelRetentionSettings - The length of time in days to retain the messages in a channel.

appInstanceRetentionSettings_channelRetentionSettings :: Lens' AppInstanceRetentionSettings (Maybe ChannelRetentionSettings)

The length of time in days to retain the messages in a channel.

AppInstanceSummary

data AppInstanceSummary

Summary of the data for an AppInstance.

See: newAppInstanceSummary smart constructor.

Constructors

AppInstanceSummary'

Fields

name :: Maybe (Sensitive Text)

The name of the AppInstance.

metadata :: Maybe (Sensitive Text)

The metadata of the AppInstance.

appInstanceArn :: Maybe Text

The AppInstance ARN.

newAppInstanceSummary :: AppInstanceSummary

Create a value of AppInstanceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:AppInstanceSummary', appInstanceSummary_name - The name of the AppInstance.

$sel:metadata:AppInstanceSummary', appInstanceSummary_metadata - The metadata of the AppInstance.

$sel:appInstanceArn:AppInstanceSummary', appInstanceSummary_appInstanceArn - The AppInstance ARN.

appInstanceSummary_name :: Lens' AppInstanceSummary (Maybe Text)

The name of the AppInstance.

appInstanceSummary_metadata :: Lens' AppInstanceSummary (Maybe Text)

The metadata of the AppInstance.

appInstanceSummary_appInstanceArn :: Lens' AppInstanceSummary (Maybe Text)

The AppInstance ARN.

AppInstanceUser

data AppInstanceUser

The details of an AppInstanceUser.

See: newAppInstanceUser smart constructor.

Constructors

AppInstanceUser'

Fields

appInstanceUserArn :: Maybe Text

The ARN of the AppInstanceUser.

name :: Maybe (Sensitive Text)

The name of the AppInstanceUser.

metadata :: Maybe (Sensitive Text)

The metadata of the AppInstanceUser.

createdTimestamp :: Maybe POSIX

The time at which the AppInstanceUser was created.

lastUpdatedTimestamp :: Maybe POSIX

The time at which the AppInstanceUser was last updated.

newAppInstanceUser :: AppInstanceUser

Create a value of AppInstanceUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUserArn:AppInstanceUser', appInstanceUser_appInstanceUserArn - The ARN of the AppInstanceUser.

$sel:name:AppInstanceUser', appInstanceUser_name - The name of the AppInstanceUser.

$sel:metadata:AppInstanceUser', appInstanceUser_metadata - The metadata of the AppInstanceUser.

$sel:createdTimestamp:AppInstanceUser', appInstanceUser_createdTimestamp - The time at which the AppInstanceUser was created.

$sel:lastUpdatedTimestamp:AppInstanceUser', appInstanceUser_lastUpdatedTimestamp - The time at which the AppInstanceUser was last updated.

appInstanceUser_appInstanceUserArn :: Lens' AppInstanceUser (Maybe Text)

The ARN of the AppInstanceUser.

appInstanceUser_name :: Lens' AppInstanceUser (Maybe Text)

The name of the AppInstanceUser.

appInstanceUser_metadata :: Lens' AppInstanceUser (Maybe Text)

The metadata of the AppInstanceUser.

appInstanceUser_createdTimestamp :: Lens' AppInstanceUser (Maybe UTCTime)

The time at which the AppInstanceUser was created.

appInstanceUser_lastUpdatedTimestamp :: Lens' AppInstanceUser (Maybe UTCTime)

The time at which the AppInstanceUser was last updated.

AppInstanceUserSummary

data AppInstanceUserSummary

Summary of the details of an AppInstanceUser.

See: newAppInstanceUserSummary smart constructor.

Constructors

AppInstanceUserSummary'

Fields

appInstanceUserArn :: Maybe Text

The ARN of the AppInstanceUser.

name :: Maybe (Sensitive Text)

The name of an AppInstanceUser.

metadata :: Maybe (Sensitive Text)

The metadata of the AppInstanceUser.

newAppInstanceUserSummary :: AppInstanceUserSummary

Create a value of AppInstanceUserSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUserArn:AppInstanceUserSummary', appInstanceUserSummary_appInstanceUserArn - The ARN of the AppInstanceUser.

$sel:name:AppInstanceUserSummary', appInstanceUserSummary_name - The name of an AppInstanceUser.

$sel:metadata:AppInstanceUserSummary', appInstanceUserSummary_metadata - The metadata of the AppInstanceUser.

appInstanceUserSummary_appInstanceUserArn :: Lens' AppInstanceUserSummary (Maybe Text)

The ARN of the AppInstanceUser.

appInstanceUserSummary_name :: Lens' AppInstanceUserSummary (Maybe Text)

The name of an AppInstanceUser.

appInstanceUserSummary_metadata :: Lens' AppInstanceUserSummary (Maybe Text)

The metadata of the AppInstanceUser.

ChannelRetentionSettings

data ChannelRetentionSettings

The details of the retention settings for a channel.

See: newChannelRetentionSettings smart constructor.

Constructors

ChannelRetentionSettings'

Fields

retentionDays :: Maybe Natural

The time in days to retain the messages in a channel.

newChannelRetentionSettings :: ChannelRetentionSettings

Create a value of ChannelRetentionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retentionDays:ChannelRetentionSettings', channelRetentionSettings_retentionDays - The time in days to retain the messages in a channel.

channelRetentionSettings_retentionDays :: Lens' ChannelRetentionSettings (Maybe Natural)

The time in days to retain the messages in a channel.

Identity

data Identity

The details of a user.

See: newIdentity smart constructor.

Constructors

Identity'

Fields

arn :: Maybe Text

The ARN in an Identity.

name :: Maybe (Sensitive Text)

The name in an Identity.

newIdentity :: Identity

Create a value of Identity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Identity', identity_arn - The ARN in an Identity.

$sel:name:Identity', identity_name - The name in an Identity.

identity_arn :: Lens' Identity (Maybe Text)

The ARN in an Identity.

identity_name :: Lens' Identity (Maybe Text)

The name in an Identity.

Tag

data Tag

Describes a tag applied to a resource.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Sensitive Text

The key of the tag.

value :: Sensitive Text

The value of the tag.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key of the tag.

$sel:value:Tag', tag_value - The value of the tag.

tag_key :: Lens' Tag Text

The key of the tag.

tag_value :: Lens' Tag Text

The value of the tag.