Network.AWS.ChimeSDKMessaging.AssociateChannelFlow

Description

Associates a channel flow with a channel. Once associated, all messages to that channel go through channel flow processors. To stop processing, use the DisassociateChannelFlow API.

Only administrators or channel moderators can associate a channel flow. The x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Creating a Request

data AssociateChannelFlow

See: newAssociateChannelFlow smart constructor.

Constructors

AssociateChannelFlow'

Fields

channelArn :: Text

The ARN of the channel.

channelFlowArn :: Text

The ARN of the channel flow.

chimeBearer :: Text

The AppInstanceUserArn of the user making the API call.

newAssociateChannelFlow

Arguments

:: Text
$sel:channelArn:AssociateChannelFlow'

-> Text
$sel:channelFlowArn:AssociateChannelFlow'

-> Text
$sel:chimeBearer:AssociateChannelFlow'

-> AssociateChannelFlow

Create a value of AssociateChannelFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:AssociateChannelFlow', associateChannelFlow_channelArn - The ARN of the channel.

$sel:channelFlowArn:AssociateChannelFlow', associateChannelFlow_channelFlowArn - The ARN of the channel flow.

$sel:chimeBearer:AssociateChannelFlow', associateChannelFlow_chimeBearer - The AppInstanceUserArn of the user making the API call.

Request Lenses

associateChannelFlow_channelArn :: Lens' AssociateChannelFlow Text

The ARN of the channel.

associateChannelFlow_channelFlowArn :: Lens' AssociateChannelFlow Text

The ARN of the channel flow.

associateChannelFlow_chimeBearer :: Lens' AssociateChannelFlow Text

The AppInstanceUserArn of the user making the API call.

Destructuring the Response

data AssociateChannelFlowResponse

See: newAssociateChannelFlowResponse smart constructor.

Constructors

AssociateChannelFlowResponse'

newAssociateChannelFlowResponse :: AssociateChannelFlowResponse

Create a value of AssociateChannelFlowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.