Network.AWS.ChimeSDKMessaging.CreateChannelFlow

Description

Creates a channel flow, a container for processors. Processors are AWS Lambda functions that perform actions on chat messages, such as stripping out profanity. You can associate channel flows with channels, and the processors in the channel flow then take action on all messages sent to that channel. This is a developer API.

Channel flows process the following items:

New and updated messages

Persistent and non-persistent messages

The Standard message type

Channel flows don't process Control or System messages. For more information about the message types provided by Chime SDK Messaging, refer to Message types in the Amazon Chime developer guide.

Creating a Request

data CreateChannelFlow

See: newCreateChannelFlow smart constructor.

Constructors

CreateChannelFlow'

Fields

tags :: Maybe (NonEmpty Tag)

The tags for the creation request.

appInstanceArn :: Text

The ARN of the channel flow request.

processors :: NonEmpty Processor

Information about the processor Lambda functions.

name :: Sensitive Text

The name of the channel flow.

clientRequestToken :: Sensitive Text

The client token for the request. An Idempotency token.

newCreateChannelFlow

Arguments

:: Text
$sel:appInstanceArn:CreateChannelFlow'

-> NonEmpty Processor
$sel:processors:CreateChannelFlow'

-> Text
$sel:name:CreateChannelFlow'

-> Text
$sel:clientRequestToken:CreateChannelFlow'

-> CreateChannelFlow

Create a value of CreateChannelFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateChannelFlow', createChannelFlow_tags - The tags for the creation request.

$sel:appInstanceArn:CreateChannelFlow', createChannelFlow_appInstanceArn - The ARN of the channel flow request.

$sel:processors:CreateChannelFlow', createChannelFlow_processors - Information about the processor Lambda functions.

$sel:name:CreateChannelFlow', createChannelFlow_name - The name of the channel flow.

$sel:clientRequestToken:CreateChannelFlow', createChannelFlow_clientRequestToken - The client token for the request. An Idempotency token.

Request Lenses

createChannelFlow_tags :: Lens' CreateChannelFlow (Maybe (NonEmpty Tag))

The tags for the creation request.

createChannelFlow_appInstanceArn :: Lens' CreateChannelFlow Text

The ARN of the channel flow request.

createChannelFlow_processors :: Lens' CreateChannelFlow (NonEmpty Processor)

Information about the processor Lambda functions.

createChannelFlow_name :: Lens' CreateChannelFlow Text

The name of the channel flow.

createChannelFlow_clientRequestToken :: Lens' CreateChannelFlow Text

The client token for the request. An Idempotency token.

Destructuring the Response

data CreateChannelFlowResponse

See: newCreateChannelFlowResponse smart constructor.

Constructors

CreateChannelFlowResponse'

Fields

channelFlowArn :: Maybe Text

The ARN of the channel flow.

httpStatus :: Int

The response's http status code.

newCreateChannelFlowResponse

Arguments

:: Int
$sel:httpStatus:CreateChannelFlowResponse'

-> CreateChannelFlowResponse

Create a value of CreateChannelFlowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelFlowArn:CreateChannelFlowResponse', createChannelFlowResponse_channelFlowArn - The ARN of the channel flow.

$sel:httpStatus:CreateChannelFlowResponse', createChannelFlowResponse_httpStatus - The response's http status code.

Response Lenses

createChannelFlowResponse_channelFlowArn :: Lens' CreateChannelFlowResponse (Maybe Text)

The ARN of the channel flow.

createChannelFlowResponse_httpStatus :: Lens' CreateChannelFlowResponse Int

The response's http status code.