Network.AWS.ChimeSDKMessaging.CreateChannelMembership

Description

Adds a user to a channel. The InvitedBy response field is derived from the request header. A channel member can:

List messages

Send messages

Receive messages

Edit their own messages

Leave the channel

Privacy settings impact this action as follows:

Public Channels: You do not need to be a member to list messages, but you must be a member to send messages.

Private Channels: You must be a member to list or send messages.

The x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Creating a Request

data CreateChannelMembership

See: newCreateChannelMembership smart constructor.

Constructors

CreateChannelMembership'

Fields

channelArn :: Text

The ARN of the channel to which you're adding users.

memberArn :: Text

The ARN of the member you want to add to the channel.

type' :: ChannelMembershipType

The membership type of a user, DEFAULT or HIDDEN. Default members are always returned as part of ListChannelMemberships. Hidden members are only returned if the type filter in ListChannelMemberships equals HIDDEN. Otherwise hidden members are not returned. This is only supported by moderators.

chimeBearer :: Text

The AppInstanceUserArn of the user that makes the API call.

newCreateChannelMembership

Arguments

:: Text
$sel:channelArn:CreateChannelMembership'

-> Text
$sel:memberArn:CreateChannelMembership'

-> ChannelMembershipType
$sel:type':CreateChannelMembership'

-> Text
$sel:chimeBearer:CreateChannelMembership'

-> CreateChannelMembership

Create a value of CreateChannelMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:CreateChannelMembership', createChannelMembership_channelArn - The ARN of the channel to which you're adding users.

$sel:memberArn:CreateChannelMembership', createChannelMembership_memberArn - The ARN of the member you want to add to the channel.

$sel:type':CreateChannelMembership', createChannelMembership_type - The membership type of a user, DEFAULT or HIDDEN. Default members are always returned as part of ListChannelMemberships. Hidden members are only returned if the type filter in ListChannelMemberships equals HIDDEN. Otherwise hidden members are not returned. This is only supported by moderators.

$sel:chimeBearer:CreateChannelMembership', createChannelMembership_chimeBearer - The AppInstanceUserArn of the user that makes the API call.

Request Lenses

createChannelMembership_channelArn :: Lens' CreateChannelMembership Text

The ARN of the channel to which you're adding users.

createChannelMembership_memberArn :: Lens' CreateChannelMembership Text

The ARN of the member you want to add to the channel.

createChannelMembership_type :: Lens' CreateChannelMembership ChannelMembershipType

The membership type of a user, DEFAULT or HIDDEN. Default members are always returned as part of ListChannelMemberships. Hidden members are only returned if the type filter in ListChannelMemberships equals HIDDEN. Otherwise hidden members are not returned. This is only supported by moderators.

createChannelMembership_chimeBearer :: Lens' CreateChannelMembership Text

The AppInstanceUserArn of the user that makes the API call.

Destructuring the Response

data CreateChannelMembershipResponse

See: newCreateChannelMembershipResponse smart constructor.

Constructors

CreateChannelMembershipResponse'

Fields

channelArn :: Maybe Text

The ARN of the channel.

member :: Maybe Identity

The ARN and metadata of the member being added.

httpStatus :: Int

The response's http status code.

newCreateChannelMembershipResponse

Arguments

:: Int
$sel:httpStatus:CreateChannelMembershipResponse'

-> CreateChannelMembershipResponse

Create a value of CreateChannelMembershipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:CreateChannelMembership', createChannelMembershipResponse_channelArn - The ARN of the channel.

$sel:member:CreateChannelMembershipResponse', createChannelMembershipResponse_member - The ARN and metadata of the member being added.

$sel:httpStatus:CreateChannelMembershipResponse', createChannelMembershipResponse_httpStatus - The response's http status code.

Response Lenses

createChannelMembershipResponse_channelArn :: Lens' CreateChannelMembershipResponse (Maybe Text)

The ARN of the channel.

createChannelMembershipResponse_member :: Lens' CreateChannelMembershipResponse (Maybe Identity)

The ARN and metadata of the member being added.

createChannelMembershipResponse_httpStatus :: Lens' CreateChannelMembershipResponse Int

The response's http status code.