Network.AWS.ChimeSDKMessaging.DisassociateChannelFlow

Description

Disassociates a channel flow from all its channels. Once disassociated, all messages to that channel stop going through the channel flow processor.

Only administrators or channel moderators can disassociate a channel flow. The x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Creating a Request

data DisassociateChannelFlow

See: newDisassociateChannelFlow smart constructor.

Constructors

DisassociateChannelFlow'

Fields

channelArn :: Text

The ARN of the channel.

channelFlowArn :: Text

The ARN of the channel flow.

chimeBearer :: Text

The AppInstanceUserArn of the user making the API call.

newDisassociateChannelFlow

Arguments

:: Text
$sel:channelArn:DisassociateChannelFlow'

-> Text
$sel:channelFlowArn:DisassociateChannelFlow'

-> Text
$sel:chimeBearer:DisassociateChannelFlow'

-> DisassociateChannelFlow

Create a value of DisassociateChannelFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:DisassociateChannelFlow', disassociateChannelFlow_channelArn - The ARN of the channel.

$sel:channelFlowArn:DisassociateChannelFlow', disassociateChannelFlow_channelFlowArn - The ARN of the channel flow.

$sel:chimeBearer:DisassociateChannelFlow', disassociateChannelFlow_chimeBearer - The AppInstanceUserArn of the user making the API call.

Request Lenses

disassociateChannelFlow_channelArn :: Lens' DisassociateChannelFlow Text

The ARN of the channel.

disassociateChannelFlow_channelFlowArn :: Lens' DisassociateChannelFlow Text

The ARN of the channel flow.

disassociateChannelFlow_chimeBearer :: Lens' DisassociateChannelFlow Text

The AppInstanceUserArn of the user making the API call.

Destructuring the Response

data DisassociateChannelFlowResponse

See: newDisassociateChannelFlowResponse smart constructor.

Constructors

DisassociateChannelFlowResponse'

newDisassociateChannelFlowResponse :: DisassociateChannelFlowResponse

Create a value of DisassociateChannelFlowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.