Network.AWS.ChimeSDKMessaging.ListChannelBans

Description

Lists all the users banned from a particular channel.

The x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Creating a Request

data ListChannelBans

See: newListChannelBans smart constructor.

Constructors

ListChannelBans'

Fields

nextToken :: Maybe (Sensitive Text)

The token passed by previous API calls until all requested bans are returned.

maxResults :: Maybe Natural

The maximum number of bans that you want returned.

channelArn :: Text

The ARN of the channel.

chimeBearer :: Text

The AppInstanceUserArn of the user that makes the API call.

newListChannelBans

Arguments

:: Text
$sel:channelArn:ListChannelBans'

-> Text
$sel:chimeBearer:ListChannelBans'

-> ListChannelBans

Create a value of ListChannelBans with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListChannelBans', listChannelBans_nextToken - The token passed by previous API calls until all requested bans are returned.

$sel:maxResults:ListChannelBans', listChannelBans_maxResults - The maximum number of bans that you want returned.

$sel:channelArn:ListChannelBans', listChannelBans_channelArn - The ARN of the channel.

$sel:chimeBearer:ListChannelBans', listChannelBans_chimeBearer - The AppInstanceUserArn of the user that makes the API call.

Request Lenses

listChannelBans_nextToken :: Lens' ListChannelBans (Maybe Text)

The token passed by previous API calls until all requested bans are returned.

listChannelBans_maxResults :: Lens' ListChannelBans (Maybe Natural)

The maximum number of bans that you want returned.

listChannelBans_channelArn :: Lens' ListChannelBans Text

The ARN of the channel.

listChannelBans_chimeBearer :: Lens' ListChannelBans Text

The AppInstanceUserArn of the user that makes the API call.

Destructuring the Response

data ListChannelBansResponse

See: newListChannelBansResponse smart constructor.

Constructors

ListChannelBansResponse'

Fields

channelArn :: Maybe Text

The ARN of the channel.

nextToken :: Maybe (Sensitive Text)

The token passed by previous API calls until all requested bans are returned.

channelBans :: Maybe [ChannelBanSummary]

The information for each requested ban.

httpStatus :: Int

The response's http status code.

newListChannelBansResponse

Arguments

:: Int
$sel:httpStatus:ListChannelBansResponse'

-> ListChannelBansResponse

Create a value of ListChannelBansResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:ListChannelBans', listChannelBansResponse_channelArn - The ARN of the channel.

$sel:nextToken:ListChannelBans', listChannelBansResponse_nextToken - The token passed by previous API calls until all requested bans are returned.

$sel:channelBans:ListChannelBansResponse', listChannelBansResponse_channelBans - The information for each requested ban.

$sel:httpStatus:ListChannelBansResponse', listChannelBansResponse_httpStatus - The response's http status code.

Response Lenses

listChannelBansResponse_channelArn :: Lens' ListChannelBansResponse (Maybe Text)

The ARN of the channel.

listChannelBansResponse_nextToken :: Lens' ListChannelBansResponse (Maybe Text)

The token passed by previous API calls until all requested bans are returned.

listChannelBansResponse_channelBans :: Lens' ListChannelBansResponse (Maybe [ChannelBanSummary])

The information for each requested ban.

listChannelBansResponse_httpStatus :: Lens' ListChannelBansResponse Int

The response's http status code.