Network.AWS.ChimeSDKMessaging.SendChannelMessage

Description

Sends a message to a particular channel that the member is a part of.

The x-amz-chime-bearer request header is mandatory. Use the AppInstanceUserArn of the user that makes the API call as the value in the header.

Also, STANDARD messages can contain 4KB of data and the 1KB of metadata. CONTROL messages can contain 30 bytes of data and no metadata.

Creating a Request

data SendChannelMessage

See: newSendChannelMessage smart constructor.

Constructors

SendChannelMessage'

Fields

metadata :: Maybe (Sensitive Text)

The optional metadata for each message.

channelArn :: Text

The ARN of the channel.

content :: Sensitive Text

The content of the message.

type' :: ChannelMessageType

The type of message, STANDARD or CONTROL.

persistence :: ChannelMessagePersistenceType

Boolean that controls whether the message is persisted on the back end. Required.

clientRequestToken :: Sensitive Text

The Idempotency token for each client request.

chimeBearer :: Text

The AppInstanceUserArn of the user that makes the API call.

newSendChannelMessage

Arguments

:: Text
$sel:channelArn:SendChannelMessage'

-> Text
$sel:content:SendChannelMessage'

-> ChannelMessageType
$sel:type':SendChannelMessage'

-> ChannelMessagePersistenceType
$sel:persistence:SendChannelMessage'

-> Text
$sel:clientRequestToken:SendChannelMessage'

-> Text
$sel:chimeBearer:SendChannelMessage'

-> SendChannelMessage

Create a value of SendChannelMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadata:SendChannelMessage', sendChannelMessage_metadata - The optional metadata for each message.

$sel:channelArn:SendChannelMessage', sendChannelMessage_channelArn - The ARN of the channel.

$sel:content:SendChannelMessage', sendChannelMessage_content - The content of the message.

$sel:type':SendChannelMessage', sendChannelMessage_type - The type of message, STANDARD or CONTROL.

$sel:persistence:SendChannelMessage', sendChannelMessage_persistence - Boolean that controls whether the message is persisted on the back end. Required.

$sel:clientRequestToken:SendChannelMessage', sendChannelMessage_clientRequestToken - The Idempotency token for each client request.

$sel:chimeBearer:SendChannelMessage', sendChannelMessage_chimeBearer - The AppInstanceUserArn of the user that makes the API call.

Request Lenses

sendChannelMessage_metadata :: Lens' SendChannelMessage (Maybe Text)

The optional metadata for each message.

sendChannelMessage_channelArn :: Lens' SendChannelMessage Text

The ARN of the channel.

sendChannelMessage_content :: Lens' SendChannelMessage Text

The content of the message.

sendChannelMessage_type :: Lens' SendChannelMessage ChannelMessageType

The type of message, STANDARD or CONTROL.

sendChannelMessage_persistence :: Lens' SendChannelMessage ChannelMessagePersistenceType

Boolean that controls whether the message is persisted on the back end. Required.

sendChannelMessage_clientRequestToken :: Lens' SendChannelMessage Text

The Idempotency token for each client request.

sendChannelMessage_chimeBearer :: Lens' SendChannelMessage Text

The AppInstanceUserArn of the user that makes the API call.

Destructuring the Response

data SendChannelMessageResponse

See: newSendChannelMessageResponse smart constructor.

Constructors

SendChannelMessageResponse'

Fields

status :: Maybe ChannelMessageStatusStructure

The status of the channel message.

channelArn :: Maybe Text

The ARN of the channel.

messageId :: Maybe Text

The ID string assigned to each message.

httpStatus :: Int

The response's http status code.

newSendChannelMessageResponse

Arguments

:: Int
$sel:httpStatus:SendChannelMessageResponse'

-> SendChannelMessageResponse

Create a value of SendChannelMessageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SendChannelMessageResponse', sendChannelMessageResponse_status - The status of the channel message.

$sel:channelArn:SendChannelMessage', sendChannelMessageResponse_channelArn - The ARN of the channel.

$sel:messageId:SendChannelMessageResponse', sendChannelMessageResponse_messageId - The ID string assigned to each message.

$sel:httpStatus:SendChannelMessageResponse', sendChannelMessageResponse_httpStatus - The response's http status code.

Response Lenses

sendChannelMessageResponse_status :: Lens' SendChannelMessageResponse (Maybe ChannelMessageStatusStructure)

The status of the channel message.

sendChannelMessageResponse_channelArn :: Lens' SendChannelMessageResponse (Maybe Text)

The ARN of the channel.

sendChannelMessageResponse_messageId :: Lens' SendChannelMessageResponse (Maybe Text)

The ID string assigned to each message.

sendChannelMessageResponse_httpStatus :: Lens' SendChannelMessageResponse Int

The response's http status code.