Network.AWS.ChimeSDKMessaging.Types

Description

Service Configuration

defaultService :: Service

API version 2021-05-15 of the Amazon Chime SDK Messaging SDK configuration.

Errors

_ThrottledClientException :: AsError a => Getting (First ServiceError) a ServiceError

The client exceeded its request rate limit.

_ResourceLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request exceeds the resource limit.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request could not be processed because of conflict in the current state of the resource.

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

The client is permanently forbidden from making the request.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the resources in the request does not exist in the system.

_ServiceFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The service encountered an unexpected error.

_UnauthorizedClientException :: AsError a => Getting (First ServiceError) a ServiceError

The client is not currently authorized to make the request.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is currently unavailable.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The input parameters don't match the service's restrictions.

ChannelMembershipType

newtype ChannelMembershipType

Constructors

ChannelMembershipType'

Fields

fromChannelMembershipType :: Text

Bundled Patterns

pattern ChannelMembershipType_DEFAULT :: ChannelMembershipType

pattern ChannelMembershipType_HIDDEN :: ChannelMembershipType

ChannelMessagePersistenceType

newtype ChannelMessagePersistenceType

Constructors

ChannelMessagePersistenceType'

Fields

fromChannelMessagePersistenceType :: Text

Bundled Patterns

pattern ChannelMessagePersistenceType_NON_PERSISTENT :: ChannelMessagePersistenceType

pattern ChannelMessagePersistenceType_PERSISTENT :: ChannelMessagePersistenceType

ChannelMessageStatus

newtype ChannelMessageStatus

Constructors

ChannelMessageStatus'

Fields

fromChannelMessageStatus :: Text

Bundled Patterns

pattern ChannelMessageStatus_DENIED :: ChannelMessageStatus

pattern ChannelMessageStatus_FAILED :: ChannelMessageStatus

pattern ChannelMessageStatus_PENDING :: ChannelMessageStatus

pattern ChannelMessageStatus_SENT :: ChannelMessageStatus

ChannelMessageType

newtype ChannelMessageType

Constructors

ChannelMessageType'

Fields

fromChannelMessageType :: Text

Bundled Patterns

pattern ChannelMessageType_CONTROL :: ChannelMessageType

pattern ChannelMessageType_STANDARD :: ChannelMessageType

ChannelMode

newtype ChannelMode

Constructors

ChannelMode'

Fields

fromChannelMode :: Text

Bundled Patterns

pattern ChannelMode_RESTRICTED :: ChannelMode

pattern ChannelMode_UNRESTRICTED :: ChannelMode

ChannelPrivacy

newtype ChannelPrivacy

Constructors

ChannelPrivacy'

Fields

fromChannelPrivacy :: Text

Bundled Patterns

pattern ChannelPrivacy_PRIVATE :: ChannelPrivacy

pattern ChannelPrivacy_PUBLIC :: ChannelPrivacy

ErrorCode

newtype ErrorCode

Constructors

ErrorCode'

Fields

fromErrorCode :: Text

Bundled Patterns

pattern ErrorCode_AccessDenied :: ErrorCode

pattern ErrorCode_BadRequest :: ErrorCode

pattern ErrorCode_Conflict :: ErrorCode

pattern ErrorCode_Forbidden :: ErrorCode

pattern ErrorCode_NotFound :: ErrorCode

pattern ErrorCode_PhoneNumberAssociationsExist :: ErrorCode

pattern ErrorCode_PreconditionFailed :: ErrorCode

pattern ErrorCode_ResourceLimitExceeded :: ErrorCode

pattern ErrorCode_ServiceFailure :: ErrorCode

pattern ErrorCode_ServiceUnavailable :: ErrorCode

pattern ErrorCode_Throttled :: ErrorCode

pattern ErrorCode_Throttling :: ErrorCode

pattern ErrorCode_Unauthorized :: ErrorCode

pattern ErrorCode_Unprocessable :: ErrorCode

pattern ErrorCode_VoiceConnectorGroupAssociationsExist :: ErrorCode

FallbackAction

newtype FallbackAction

Constructors

FallbackAction'

Fields

fromFallbackAction :: Text

Bundled Patterns

pattern FallbackAction_ABORT :: FallbackAction

pattern FallbackAction_CONTINUE :: FallbackAction

InvocationType

newtype InvocationType

Constructors

InvocationType'

Fields

fromInvocationType :: Text

Bundled Patterns

pattern InvocationType_ASYNC :: InvocationType

SortOrder

newtype SortOrder

Constructors

SortOrder'

Fields

fromSortOrder :: Text

Bundled Patterns

pattern SortOrder_ASCENDING :: SortOrder

pattern SortOrder_DESCENDING :: SortOrder

AppInstanceUserMembershipSummary

data AppInstanceUserMembershipSummary

Summary of the membership details of an AppInstanceUser.

See: newAppInstanceUserMembershipSummary smart constructor.

Constructors

AppInstanceUserMembershipSummary'

Fields

readMarkerTimestamp :: Maybe POSIX

The time at which a message was last read.

type' :: Maybe ChannelMembershipType

The type of ChannelMembership.

newAppInstanceUserMembershipSummary :: AppInstanceUserMembershipSummary

Create a value of AppInstanceUserMembershipSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readMarkerTimestamp:AppInstanceUserMembershipSummary', appInstanceUserMembershipSummary_readMarkerTimestamp - The time at which a message was last read.

$sel:type':AppInstanceUserMembershipSummary', appInstanceUserMembershipSummary_type - The type of ChannelMembership.

appInstanceUserMembershipSummary_readMarkerTimestamp :: Lens' AppInstanceUserMembershipSummary (Maybe UTCTime)

The time at which a message was last read.

appInstanceUserMembershipSummary_type :: Lens' AppInstanceUserMembershipSummary (Maybe ChannelMembershipType)

The type of ChannelMembership.

BatchChannelMemberships

data BatchChannelMemberships

The membership information, including member ARNs, the channel ARN, and membership types.

See: newBatchChannelMemberships smart constructor.

Constructors

BatchChannelMemberships'

Fields

members :: Maybe [Identity]

The users successfully added to the request.

channelArn :: Maybe Text

The ARN of the channel to which you're adding users.

type' :: Maybe ChannelMembershipType

The membership types set for the channel users.

invitedBy :: Maybe Identity

The identifier of the member who invited another member.

newBatchChannelMemberships :: BatchChannelMemberships

Create a value of BatchChannelMemberships with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:members:BatchChannelMemberships', batchChannelMemberships_members - The users successfully added to the request.

$sel:channelArn:BatchChannelMemberships', batchChannelMemberships_channelArn - The ARN of the channel to which you're adding users.

$sel:type':BatchChannelMemberships', batchChannelMemberships_type - The membership types set for the channel users.

$sel:invitedBy:BatchChannelMemberships', batchChannelMemberships_invitedBy - The identifier of the member who invited another member.

batchChannelMemberships_members :: Lens' BatchChannelMemberships (Maybe [Identity])

The users successfully added to the request.

batchChannelMemberships_channelArn :: Lens' BatchChannelMemberships (Maybe Text)

The ARN of the channel to which you're adding users.

batchChannelMemberships_type :: Lens' BatchChannelMemberships (Maybe ChannelMembershipType)

The membership types set for the channel users.

batchChannelMemberships_invitedBy :: Lens' BatchChannelMemberships (Maybe Identity)

The identifier of the member who invited another member.

BatchCreateChannelMembershipError

data BatchCreateChannelMembershipError

A list of failed member ARNs, error codes, and error messages.

See: newBatchCreateChannelMembershipError smart constructor.

Constructors

BatchCreateChannelMembershipError'

Fields

errorCode :: Maybe ErrorCode

The error code.

memberArn :: Maybe Text

The ARN of the member that the service couldn't add.

errorMessage :: Maybe Text

The error message.

newBatchCreateChannelMembershipError :: BatchCreateChannelMembershipError

Create a value of BatchCreateChannelMembershipError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:BatchCreateChannelMembershipError', batchCreateChannelMembershipError_errorCode - The error code.

$sel:memberArn:BatchCreateChannelMembershipError', batchCreateChannelMembershipError_memberArn - The ARN of the member that the service couldn't add.

$sel:errorMessage:BatchCreateChannelMembershipError', batchCreateChannelMembershipError_errorMessage - The error message.

batchCreateChannelMembershipError_errorCode :: Lens' BatchCreateChannelMembershipError (Maybe ErrorCode)

The error code.

batchCreateChannelMembershipError_memberArn :: Lens' BatchCreateChannelMembershipError (Maybe Text)

The ARN of the member that the service couldn't add.

batchCreateChannelMembershipError_errorMessage :: Lens' BatchCreateChannelMembershipError (Maybe Text)

The error message.

Channel

data Channel

The details of a channel.

See: newChannel smart constructor.

Constructors

Channel'

Fields

mode :: Maybe ChannelMode

The mode of the channel.

createdBy :: Maybe Identity

The AppInstanceUser who created the channel.

channelArn :: Maybe Text

The ARN of a channel.

privacy :: Maybe ChannelPrivacy

The channel's privacy setting.

channelFlowArn :: Maybe Text

The ARN of the channel flow.

lastMessageTimestamp :: Maybe POSIX

The time at which a member sent the last message in the channel.

name :: Maybe (Sensitive Text)

The name of a channel.

metadata :: Maybe (Sensitive Text)

The channel's metadata.

createdTimestamp :: Maybe POSIX

The time at which the AppInstanceUser created the channel.

lastUpdatedTimestamp :: Maybe POSIX

The time at which a channel was last updated.

newChannel :: Channel

Create a value of Channel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mode:Channel', channel_mode - The mode of the channel.

$sel:createdBy:Channel', channel_createdBy - The AppInstanceUser who created the channel.

$sel:channelArn:Channel', channel_channelArn - The ARN of a channel.

$sel:privacy:Channel', channel_privacy - The channel's privacy setting.

$sel:channelFlowArn:Channel', channel_channelFlowArn - The ARN of the channel flow.

$sel:lastMessageTimestamp:Channel', channel_lastMessageTimestamp - The time at which a member sent the last message in the channel.

$sel:name:Channel', channel_name - The name of a channel.

$sel:metadata:Channel', channel_metadata - The channel's metadata.

$sel:createdTimestamp:Channel', channel_createdTimestamp - The time at which the AppInstanceUser created the channel.

$sel:lastUpdatedTimestamp:Channel', channel_lastUpdatedTimestamp - The time at which a channel was last updated.

channel_mode :: Lens' Channel (Maybe ChannelMode)

The mode of the channel.

channel_createdBy :: Lens' Channel (Maybe Identity)

The AppInstanceUser who created the channel.

channel_channelArn :: Lens' Channel (Maybe Text)

The ARN of a channel.

channel_privacy :: Lens' Channel (Maybe ChannelPrivacy)

The channel's privacy setting.

channel_channelFlowArn :: Lens' Channel (Maybe Text)

The ARN of the channel flow.

channel_lastMessageTimestamp :: Lens' Channel (Maybe UTCTime)

The time at which a member sent the last message in the channel.

channel_name :: Lens' Channel (Maybe Text)

The name of a channel.

channel_metadata :: Lens' Channel (Maybe Text)

The channel's metadata.

channel_createdTimestamp :: Lens' Channel (Maybe UTCTime)

The time at which the AppInstanceUser created the channel.

channel_lastUpdatedTimestamp :: Lens' Channel (Maybe UTCTime)

The time at which a channel was last updated.

ChannelAssociatedWithFlowSummary

data ChannelAssociatedWithFlowSummary

Summary of details of a channel associated with channel flow.

See: newChannelAssociatedWithFlowSummary smart constructor.

Constructors

ChannelAssociatedWithFlowSummary'

Fields

mode :: Maybe ChannelMode

The mode of the channel.

channelArn :: Maybe Text

The ARN of the channel.

privacy :: Maybe ChannelPrivacy

The channel's privacy setting.

name :: Maybe (Sensitive Text)

The name of the channel flow.

metadata :: Maybe (Sensitive Text)

The channel's metadata.

newChannelAssociatedWithFlowSummary :: ChannelAssociatedWithFlowSummary

Create a value of ChannelAssociatedWithFlowSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mode:ChannelAssociatedWithFlowSummary', channelAssociatedWithFlowSummary_mode - The mode of the channel.

$sel:channelArn:ChannelAssociatedWithFlowSummary', channelAssociatedWithFlowSummary_channelArn - The ARN of the channel.

$sel:privacy:ChannelAssociatedWithFlowSummary', channelAssociatedWithFlowSummary_privacy - The channel's privacy setting.

$sel:name:ChannelAssociatedWithFlowSummary', channelAssociatedWithFlowSummary_name - The name of the channel flow.

$sel:metadata:ChannelAssociatedWithFlowSummary', channelAssociatedWithFlowSummary_metadata - The channel's metadata.

channelAssociatedWithFlowSummary_mode :: Lens' ChannelAssociatedWithFlowSummary (Maybe ChannelMode)

The mode of the channel.

channelAssociatedWithFlowSummary_channelArn :: Lens' ChannelAssociatedWithFlowSummary (Maybe Text)

The ARN of the channel.

channelAssociatedWithFlowSummary_privacy :: Lens' ChannelAssociatedWithFlowSummary (Maybe ChannelPrivacy)

The channel's privacy setting.

channelAssociatedWithFlowSummary_name :: Lens' ChannelAssociatedWithFlowSummary (Maybe Text)

The name of the channel flow.

channelAssociatedWithFlowSummary_metadata :: Lens' ChannelAssociatedWithFlowSummary (Maybe Text)

The channel's metadata.

ChannelBan

data ChannelBan

The details of a channel ban.

See: newChannelBan smart constructor.

Constructors

ChannelBan'

Fields

createdBy :: Maybe Identity

The AppInstanceUser who created the ban.

channelArn :: Maybe Text

The ARN of the channel from which a member is being banned.

member :: Maybe Identity

The member being banned from the channel.

createdTimestamp :: Maybe POSIX

The time at which the ban was created.

newChannelBan :: ChannelBan

Create a value of ChannelBan with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdBy:ChannelBan', channelBan_createdBy - The AppInstanceUser who created the ban.

$sel:channelArn:ChannelBan', channelBan_channelArn - The ARN of the channel from which a member is being banned.

$sel:member:ChannelBan', channelBan_member - The member being banned from the channel.

$sel:createdTimestamp:ChannelBan', channelBan_createdTimestamp - The time at which the ban was created.

channelBan_createdBy :: Lens' ChannelBan (Maybe Identity)

The AppInstanceUser who created the ban.

channelBan_channelArn :: Lens' ChannelBan (Maybe Text)

The ARN of the channel from which a member is being banned.

channelBan_member :: Lens' ChannelBan (Maybe Identity)

The member being banned from the channel.

channelBan_createdTimestamp :: Lens' ChannelBan (Maybe UTCTime)

The time at which the ban was created.

ChannelBanSummary

data ChannelBanSummary

Summary of the details of a ChannelBan.

See: newChannelBanSummary smart constructor.

Constructors

ChannelBanSummary'

Fields

member :: Maybe Identity

The member being banned from a channel.

newChannelBanSummary :: ChannelBanSummary

Create a value of ChannelBanSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:member:ChannelBanSummary', channelBanSummary_member - The member being banned from a channel.

channelBanSummary_member :: Lens' ChannelBanSummary (Maybe Identity)

The member being banned from a channel.

ChannelFlow

data ChannelFlow

The details of a channel flow.

See: newChannelFlow smart constructor.

Constructors

ChannelFlow'

Fields

processors :: Maybe (NonEmpty Processor)

Information about the processor Lambda functions.

channelFlowArn :: Maybe Text

The ARN of the channel flow.

name :: Maybe (Sensitive Text)

The name of the channel flow.

createdTimestamp :: Maybe POSIX

The time at which the channel flow was created.

lastUpdatedTimestamp :: Maybe POSIX

The time at which a channel flow was updated.

newChannelFlow :: ChannelFlow

Create a value of ChannelFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:processors:ChannelFlow', channelFlow_processors - Information about the processor Lambda functions.

$sel:channelFlowArn:ChannelFlow', channelFlow_channelFlowArn - The ARN of the channel flow.

$sel:name:ChannelFlow', channelFlow_name - The name of the channel flow.

$sel:createdTimestamp:ChannelFlow', channelFlow_createdTimestamp - The time at which the channel flow was created.

$sel:lastUpdatedTimestamp:ChannelFlow', channelFlow_lastUpdatedTimestamp - The time at which a channel flow was updated.

channelFlow_processors :: Lens' ChannelFlow (Maybe (NonEmpty Processor))

Information about the processor Lambda functions.

channelFlow_channelFlowArn :: Lens' ChannelFlow (Maybe Text)

The ARN of the channel flow.

channelFlow_name :: Lens' ChannelFlow (Maybe Text)

The name of the channel flow.

channelFlow_createdTimestamp :: Lens' ChannelFlow (Maybe UTCTime)

The time at which the channel flow was created.

channelFlow_lastUpdatedTimestamp :: Lens' ChannelFlow (Maybe UTCTime)

The time at which a channel flow was updated.

ChannelFlowSummary

data ChannelFlowSummary

Summary of details of a channel flow.

See: newChannelFlowSummary smart constructor.

Constructors

ChannelFlowSummary'

Fields

processors :: Maybe (NonEmpty Processor)

Information about the processor Lambda functions.

channelFlowArn :: Maybe Text

The ARN of the channel flow.

name :: Maybe (Sensitive Text)

The name of the channel flow.

newChannelFlowSummary :: ChannelFlowSummary

Create a value of ChannelFlowSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:processors:ChannelFlowSummary', channelFlowSummary_processors - Information about the processor Lambda functions.

$sel:channelFlowArn:ChannelFlowSummary', channelFlowSummary_channelFlowArn - The ARN of the channel flow.

$sel:name:ChannelFlowSummary', channelFlowSummary_name - The name of the channel flow.

channelFlowSummary_processors :: Lens' ChannelFlowSummary (Maybe (NonEmpty Processor))

Information about the processor Lambda functions.

channelFlowSummary_channelFlowArn :: Lens' ChannelFlowSummary (Maybe Text)

The ARN of the channel flow.

channelFlowSummary_name :: Lens' ChannelFlowSummary (Maybe Text)

The name of the channel flow.

ChannelMembership

data ChannelMembership

The details of a channel member.

See: newChannelMembership smart constructor.

Constructors

ChannelMembership'

Fields

channelArn :: Maybe Text

The ARN of the member's channel.

member :: Maybe Identity

The data of the channel member.

type' :: Maybe ChannelMembershipType

The membership type set for the channel member.

invitedBy :: Maybe Identity

The identifier of the member who invited another member.

createdTimestamp :: Maybe POSIX

The time at which the channel membership was created.

lastUpdatedTimestamp :: Maybe POSIX

The time at which a channel membership was last updated.

newChannelMembership :: ChannelMembership

Create a value of ChannelMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelArn:ChannelMembership', channelMembership_channelArn - The ARN of the member's channel.

$sel:member:ChannelMembership', channelMembership_member - The data of the channel member.

$sel:type':ChannelMembership', channelMembership_type - The membership type set for the channel member.

$sel:invitedBy:ChannelMembership', channelMembership_invitedBy - The identifier of the member who invited another member.

$sel:createdTimestamp:ChannelMembership', channelMembership_createdTimestamp - The time at which the channel membership was created.

$sel:lastUpdatedTimestamp:ChannelMembership', channelMembership_lastUpdatedTimestamp - The time at which a channel membership was last updated.

channelMembership_channelArn :: Lens' ChannelMembership (Maybe Text)

The ARN of the member's channel.

channelMembership_member :: Lens' ChannelMembership (Maybe Identity)

The data of the channel member.

channelMembership_type :: Lens' ChannelMembership (Maybe ChannelMembershipType)

The membership type set for the channel member.

channelMembership_invitedBy :: Lens' ChannelMembership (Maybe Identity)

The identifier of the member who invited another member.

channelMembership_createdTimestamp :: Lens' ChannelMembership (Maybe UTCTime)

The time at which the channel membership was created.

channelMembership_lastUpdatedTimestamp :: Lens' ChannelMembership (Maybe UTCTime)

The time at which a channel membership was last updated.

ChannelMembershipForAppInstanceUserSummary

data ChannelMembershipForAppInstanceUserSummary

Summary of the channel membership details of an AppInstanceUser.

See: newChannelMembershipForAppInstanceUserSummary smart constructor.

Constructors

ChannelMembershipForAppInstanceUserSummary'

Fields

appInstanceUserMembershipSummary :: Maybe AppInstanceUserMembershipSummary

Returns the channel membership data for an AppInstance.

channelSummary :: Maybe ChannelSummary

Returns the channel data for an AppInstance.

newChannelMembershipForAppInstanceUserSummary :: ChannelMembershipForAppInstanceUserSummary

Create a value of ChannelMembershipForAppInstanceUserSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appInstanceUserMembershipSummary:ChannelMembershipForAppInstanceUserSummary', channelMembershipForAppInstanceUserSummary_appInstanceUserMembershipSummary - Returns the channel membership data for an AppInstance.

$sel:channelSummary:ChannelMembershipForAppInstanceUserSummary', channelMembershipForAppInstanceUserSummary_channelSummary - Returns the channel data for an AppInstance.

channelMembershipForAppInstanceUserSummary_appInstanceUserMembershipSummary :: Lens' ChannelMembershipForAppInstanceUserSummary (Maybe AppInstanceUserMembershipSummary)

Returns the channel membership data for an AppInstance.

channelMembershipForAppInstanceUserSummary_channelSummary :: Lens' ChannelMembershipForAppInstanceUserSummary (Maybe ChannelSummary)

Returns the channel data for an AppInstance.

ChannelMembershipSummary

data ChannelMembershipSummary

Summary of the details of a ChannelMembership.

See: newChannelMembershipSummary smart constructor.

Constructors

ChannelMembershipSummary'

Fields

member :: Maybe Identity

A member's summary data.

newChannelMembershipSummary :: ChannelMembershipSummary

Create a value of ChannelMembershipSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:member:ChannelMembershipSummary', channelMembershipSummary_member - A member's summary data.

channelMembershipSummary_member :: Lens' ChannelMembershipSummary (Maybe Identity)

A member's summary data.

ChannelMessage

data ChannelMessage

The details of a message in a channel.

See: newChannelMessage smart constructor.

Constructors

ChannelMessage'

Fields

status :: Maybe ChannelMessageStatusStructure

The status of the channel message.

sender :: Maybe Identity

The message sender.

channelArn :: Maybe Text

The ARN of the channel.

content :: Maybe (Sensitive Text)

The message content.

redacted :: Maybe Bool

Hides the content of a message.

persistence :: Maybe ChannelMessagePersistenceType

The persistence setting for a channel message.

metadata :: Maybe (Sensitive Text)

The message metadata.

type' :: Maybe ChannelMessageType

The message type.

createdTimestamp :: Maybe POSIX

The time at which the message was created.

messageId :: Maybe Text

The ID of a message.

lastUpdatedTimestamp :: Maybe POSIX

The time at which a message was updated.

lastEditedTimestamp :: Maybe POSIX

The time at which a message was edited.

newChannelMessage :: ChannelMessage

Create a value of ChannelMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ChannelMessage', channelMessage_status - The status of the channel message.

$sel:sender:ChannelMessage', channelMessage_sender - The message sender.

$sel:channelArn:ChannelMessage', channelMessage_channelArn - The ARN of the channel.

$sel:content:ChannelMessage', channelMessage_content - The message content.

$sel:redacted:ChannelMessage', channelMessage_redacted - Hides the content of a message.

$sel:persistence:ChannelMessage', channelMessage_persistence - The persistence setting for a channel message.

$sel:metadata:ChannelMessage', channelMessage_metadata - The message metadata.

$sel:type':ChannelMessage', channelMessage_type - The message type.

$sel:createdTimestamp:ChannelMessage', channelMessage_createdTimestamp - The time at which the message was created.

$sel:messageId:ChannelMessage', channelMessage_messageId - The ID of a message.

$sel:lastUpdatedTimestamp:ChannelMessage', channelMessage_lastUpdatedTimestamp - The time at which a message was updated.

$sel:lastEditedTimestamp:ChannelMessage', channelMessage_lastEditedTimestamp - The time at which a message was edited.

channelMessage_status :: Lens' ChannelMessage (Maybe ChannelMessageStatusStructure)

The status of the channel message.

channelMessage_sender :: Lens' ChannelMessage (Maybe Identity)

The message sender.

channelMessage_channelArn :: Lens' ChannelMessage (Maybe Text)

The ARN of the channel.

channelMessage_content :: Lens' ChannelMessage (Maybe Text)

The message content.

channelMessage_redacted :: Lens' ChannelMessage (Maybe Bool)

Hides the content of a message.

channelMessage_persistence :: Lens' ChannelMessage (Maybe ChannelMessagePersistenceType)

The persistence setting for a channel message.

channelMessage_metadata :: Lens' ChannelMessage (Maybe Text)

The message metadata.

channelMessage_type :: Lens' ChannelMessage (Maybe ChannelMessageType)

The message type.

channelMessage_createdTimestamp :: Lens' ChannelMessage (Maybe UTCTime)

The time at which the message was created.

channelMessage_messageId :: Lens' ChannelMessage (Maybe Text)

The ID of a message.

channelMessage_lastUpdatedTimestamp :: Lens' ChannelMessage (Maybe UTCTime)

The time at which a message was updated.

channelMessage_lastEditedTimestamp :: Lens' ChannelMessage (Maybe UTCTime)

The time at which a message was edited.

ChannelMessageCallback

data ChannelMessageCallback

Stores information about a callback.

See: newChannelMessageCallback smart constructor.

Constructors

ChannelMessageCallback'

Fields

content :: Maybe (Sensitive Text)

The message content.

metadata :: Maybe (Sensitive Text)

The message metadata.

messageId :: Text

The message ID.

newChannelMessageCallback

Arguments

:: Text
$sel:messageId:ChannelMessageCallback'

-> ChannelMessageCallback

Create a value of ChannelMessageCallback with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:content:ChannelMessageCallback', channelMessageCallback_content - The message content.

$sel:metadata:ChannelMessageCallback', channelMessageCallback_metadata - The message metadata.

$sel:messageId:ChannelMessageCallback', channelMessageCallback_messageId - The message ID.

channelMessageCallback_content :: Lens' ChannelMessageCallback (Maybe Text)

The message content.

channelMessageCallback_metadata :: Lens' ChannelMessageCallback (Maybe Text)

The message metadata.

channelMessageCallback_messageId :: Lens' ChannelMessageCallback Text

The message ID.

ChannelMessageStatusStructure

data ChannelMessageStatusStructure

Stores information about a message status.

See: newChannelMessageStatusStructure smart constructor.

Constructors

ChannelMessageStatusStructure'

Fields

value :: Maybe ChannelMessageStatus

The message status value.

detail :: Maybe Text

Contains more details about the messasge status.

newChannelMessageStatusStructure :: ChannelMessageStatusStructure

Create a value of ChannelMessageStatusStructure with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:ChannelMessageStatusStructure', channelMessageStatusStructure_value - The message status value.

$sel:detail:ChannelMessageStatusStructure', channelMessageStatusStructure_detail - Contains more details about the messasge status.

channelMessageStatusStructure_value :: Lens' ChannelMessageStatusStructure (Maybe ChannelMessageStatus)

The message status value.

channelMessageStatusStructure_detail :: Lens' ChannelMessageStatusStructure (Maybe Text)

Contains more details about the messasge status.

ChannelMessageSummary

data ChannelMessageSummary

Summary of the messages in a Channel.

See: newChannelMessageSummary smart constructor.

Constructors

ChannelMessageSummary'

Fields

status :: Maybe ChannelMessageStatusStructure

The message status. The status value is SENT for messages sent to a channel without a channel flow. For channels associated with channel flow, the value determines the processing stage.

sender :: Maybe Identity

The message sender.

content :: Maybe (Sensitive Text)

The content of the message.

redacted :: Maybe Bool

Indicates whether a message was redacted.

metadata :: Maybe (Sensitive Text)

The metadata of the message.

type' :: Maybe ChannelMessageType

The type of message.

createdTimestamp :: Maybe POSIX

The time at which the message summary was created.

messageId :: Maybe Text

The ID of the message.

lastUpdatedTimestamp :: Maybe POSIX

The time at which a message was last updated.

lastEditedTimestamp :: Maybe POSIX

The time at which a message was last edited.

newChannelMessageSummary :: ChannelMessageSummary

Create a value of ChannelMessageSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ChannelMessageSummary', channelMessageSummary_status - The message status. The status value is SENT for messages sent to a channel without a channel flow. For channels associated with channel flow, the value determines the processing stage.

$sel:sender:ChannelMessageSummary', channelMessageSummary_sender - The message sender.

$sel:content:ChannelMessageSummary', channelMessageSummary_content - The content of the message.

$sel:redacted:ChannelMessageSummary', channelMessageSummary_redacted - Indicates whether a message was redacted.

$sel:metadata:ChannelMessageSummary', channelMessageSummary_metadata - The metadata of the message.

$sel:type':ChannelMessageSummary', channelMessageSummary_type - The type of message.

$sel:createdTimestamp:ChannelMessageSummary', channelMessageSummary_createdTimestamp - The time at which the message summary was created.

$sel:messageId:ChannelMessageSummary', channelMessageSummary_messageId - The ID of the message.

$sel:lastUpdatedTimestamp:ChannelMessageSummary', channelMessageSummary_lastUpdatedTimestamp - The time at which a message was last updated.

$sel:lastEditedTimestamp:ChannelMessageSummary', channelMessageSummary_lastEditedTimestamp - The time at which a message was last edited.

channelMessageSummary_status :: Lens' ChannelMessageSummary (Maybe ChannelMessageStatusStructure)

The message status. The status value is SENT for messages sent to a channel without a channel flow. For channels associated with channel flow, the value determines the processing stage.

channelMessageSummary_sender :: Lens' ChannelMessageSummary (Maybe Identity)

The message sender.

channelMessageSummary_content :: Lens' ChannelMessageSummary (Maybe Text)

The content of the message.

channelMessageSummary_redacted :: Lens' ChannelMessageSummary (Maybe Bool)

Indicates whether a message was redacted.

channelMessageSummary_metadata :: Lens' ChannelMessageSummary (Maybe Text)

The metadata of the message.

channelMessageSummary_type :: Lens' ChannelMessageSummary (Maybe ChannelMessageType)

The type of message.

channelMessageSummary_createdTimestamp :: Lens' ChannelMessageSummary (Maybe UTCTime)

The time at which the message summary was created.

channelMessageSummary_messageId :: Lens' ChannelMessageSummary (Maybe Text)

The ID of the message.

channelMessageSummary_lastUpdatedTimestamp :: Lens' ChannelMessageSummary (Maybe UTCTime)

The time at which a message was last updated.

channelMessageSummary_lastEditedTimestamp :: Lens' ChannelMessageSummary (Maybe UTCTime)

The time at which a message was last edited.

ChannelModeratedByAppInstanceUserSummary

data ChannelModeratedByAppInstanceUserSummary

Summary of the details of a moderated channel.

See: newChannelModeratedByAppInstanceUserSummary smart constructor.

Constructors

ChannelModeratedByAppInstanceUserSummary'

Fields

channelSummary :: Maybe ChannelSummary

Summary of the details of a Channel.

newChannelModeratedByAppInstanceUserSummary :: ChannelModeratedByAppInstanceUserSummary

Create a value of ChannelModeratedByAppInstanceUserSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelSummary:ChannelModeratedByAppInstanceUserSummary', channelModeratedByAppInstanceUserSummary_channelSummary - Summary of the details of a Channel.

channelModeratedByAppInstanceUserSummary_channelSummary :: Lens' ChannelModeratedByAppInstanceUserSummary (Maybe ChannelSummary)

Summary of the details of a Channel.

ChannelModerator

data ChannelModerator

The details of a channel moderator.

See: newChannelModerator smart constructor.

Constructors

ChannelModerator'

Fields

createdBy :: Maybe Identity

The AppInstanceUser who created the moderator.

channelArn :: Maybe Text

The ARN of the moderator's channel.

createdTimestamp :: Maybe POSIX

The time at which the moderator was created.

moderator :: Maybe Identity

The moderator's data.

newChannelModerator :: ChannelModerator

Create a value of ChannelModerator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdBy:ChannelModerator', channelModerator_createdBy - The AppInstanceUser who created the moderator.

$sel:channelArn:ChannelModerator', channelModerator_channelArn - The ARN of the moderator's channel.

$sel:createdTimestamp:ChannelModerator', channelModerator_createdTimestamp - The time at which the moderator was created.

$sel:moderator:ChannelModerator', channelModerator_moderator - The moderator's data.

channelModerator_createdBy :: Lens' ChannelModerator (Maybe Identity)

The AppInstanceUser who created the moderator.

channelModerator_channelArn :: Lens' ChannelModerator (Maybe Text)

The ARN of the moderator's channel.

channelModerator_createdTimestamp :: Lens' ChannelModerator (Maybe UTCTime)

The time at which the moderator was created.

channelModerator_moderator :: Lens' ChannelModerator (Maybe Identity)

The moderator's data.

ChannelModeratorSummary

data ChannelModeratorSummary

Summary of the details of a ChannelModerator.

See: newChannelModeratorSummary smart constructor.

Constructors

ChannelModeratorSummary'

Fields

moderator :: Maybe Identity

The data for a moderator.

newChannelModeratorSummary :: ChannelModeratorSummary

Create a value of ChannelModeratorSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:moderator:ChannelModeratorSummary', channelModeratorSummary_moderator - The data for a moderator.

channelModeratorSummary_moderator :: Lens' ChannelModeratorSummary (Maybe Identity)

The data for a moderator.

ChannelSummary

data ChannelSummary

Summary of the details of a Channel.

See: newChannelSummary smart constructor.

Constructors

ChannelSummary'

Fields

mode :: Maybe ChannelMode

The mode of the channel.

channelArn :: Maybe Text

The ARN of the channel.

privacy :: Maybe ChannelPrivacy

The privacy setting of the channel.

lastMessageTimestamp :: Maybe POSIX

The time at which the last message in a channel was sent.

name :: Maybe (Sensitive Text)

The name of the channel.

metadata :: Maybe (Sensitive Text)

The metadata of the channel.

newChannelSummary :: ChannelSummary

Create a value of ChannelSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mode:ChannelSummary', channelSummary_mode - The mode of the channel.

$sel:channelArn:ChannelSummary', channelSummary_channelArn - The ARN of the channel.

$sel:privacy:ChannelSummary', channelSummary_privacy - The privacy setting of the channel.

$sel:lastMessageTimestamp:ChannelSummary', channelSummary_lastMessageTimestamp - The time at which the last message in a channel was sent.

$sel:name:ChannelSummary', channelSummary_name - The name of the channel.

$sel:metadata:ChannelSummary', channelSummary_metadata - The metadata of the channel.

channelSummary_mode :: Lens' ChannelSummary (Maybe ChannelMode)

The mode of the channel.

channelSummary_channelArn :: Lens' ChannelSummary (Maybe Text)

The ARN of the channel.

channelSummary_privacy :: Lens' ChannelSummary (Maybe ChannelPrivacy)

The privacy setting of the channel.

channelSummary_lastMessageTimestamp :: Lens' ChannelSummary (Maybe UTCTime)

The time at which the last message in a channel was sent.

channelSummary_name :: Lens' ChannelSummary (Maybe Text)

The name of the channel.

channelSummary_metadata :: Lens' ChannelSummary (Maybe Text)

The metadata of the channel.

Identity

data Identity

The details of a user.

See: newIdentity smart constructor.

Constructors

Identity'

Fields

arn :: Maybe Text

The ARN in an Identity.

name :: Maybe (Sensitive Text)

The name in an Identity.

newIdentity :: Identity

Create a value of Identity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Identity', identity_arn - The ARN in an Identity.

$sel:name:Identity', identity_name - The name in an Identity.

identity_arn :: Lens' Identity (Maybe Text)

The ARN in an Identity.

identity_name :: Lens' Identity (Maybe Text)

The name in an Identity.

LambdaConfiguration

data LambdaConfiguration

Stores metadata about a Lambda processor.

See: newLambdaConfiguration smart constructor.

Constructors

LambdaConfiguration'

Fields

resourceArn :: Text

The ARN of the Lambda message processing function.

invocationType :: InvocationType

Controls how the Lambda function is invoked.

newLambdaConfiguration

Arguments

:: Text
$sel:resourceArn:LambdaConfiguration'

-> InvocationType
$sel:invocationType:LambdaConfiguration'

-> LambdaConfiguration

Create a value of LambdaConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:LambdaConfiguration', lambdaConfiguration_resourceArn - The ARN of the Lambda message processing function.

$sel:invocationType:LambdaConfiguration', lambdaConfiguration_invocationType - Controls how the Lambda function is invoked.

lambdaConfiguration_resourceArn :: Lens' LambdaConfiguration Text

The ARN of the Lambda message processing function.

lambdaConfiguration_invocationType :: Lens' LambdaConfiguration InvocationType

Controls how the Lambda function is invoked.

MessagingSessionEndpoint

data MessagingSessionEndpoint

The websocket endpoint used to connect to Amazon Chime SDK messaging.

See: newMessagingSessionEndpoint smart constructor.

Constructors

MessagingSessionEndpoint'

Fields

url :: Maybe Text

The endpoint to which you establish a websocket connection.

newMessagingSessionEndpoint :: MessagingSessionEndpoint

Create a value of MessagingSessionEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:MessagingSessionEndpoint', messagingSessionEndpoint_url - The endpoint to which you establish a websocket connection.

messagingSessionEndpoint_url :: Lens' MessagingSessionEndpoint (Maybe Text)

The endpoint to which you establish a websocket connection.

Processor

data Processor

The information about a processor in a channel flow.

See: newProcessor smart constructor.

Constructors

Processor'

Fields

name :: Sensitive Text

The name of the channel flow.

configuration :: ProcessorConfiguration

The information about the type of processor and its identifier.

executionOrder :: Natural

The sequence in which processors run. If you have multiple processors in a channel flow, message processing goes through each processor in the sequence. The value determines the sequence. At this point, we support only 1 processor within a flow.

fallbackAction :: FallbackAction

Determines whether to continue or stop processing if communication with processor fails. If the last processor in a channel flow sequence has a fallback action of CONTINUE, and communication with the processor fails, the message is considered processed and sent to the recipients in the channel.

newProcessor

Arguments

:: Text
$sel:name:Processor'

-> ProcessorConfiguration
$sel:configuration:Processor'

-> Natural
$sel:executionOrder:Processor'

-> FallbackAction
$sel:fallbackAction:Processor'

-> Processor

Create a value of Processor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Processor', processor_name - The name of the channel flow.

$sel:configuration:Processor', processor_configuration - The information about the type of processor and its identifier.

$sel:executionOrder:Processor', processor_executionOrder - The sequence in which processors run. If you have multiple processors in a channel flow, message processing goes through each processor in the sequence. The value determines the sequence. At this point, we support only 1 processor within a flow.

$sel:fallbackAction:Processor', processor_fallbackAction - Determines whether to continue or stop processing if communication with processor fails. If the last processor in a channel flow sequence has a fallback action of CONTINUE, and communication with the processor fails, the message is considered processed and sent to the recipients in the channel.

processor_name :: Lens' Processor Text

The name of the channel flow.

processor_configuration :: Lens' Processor ProcessorConfiguration

The information about the type of processor and its identifier.

processor_executionOrder :: Lens' Processor Natural

The sequence in which processors run. If you have multiple processors in a channel flow, message processing goes through each processor in the sequence. The value determines the sequence. At this point, we support only 1 processor within a flow.

processor_fallbackAction :: Lens' Processor FallbackAction

Determines whether to continue or stop processing if communication with processor fails. If the last processor in a channel flow sequence has a fallback action of CONTINUE, and communication with the processor fails, the message is considered processed and sent to the recipients in the channel.

ProcessorConfiguration

data ProcessorConfiguration

A processor's metadata.

See: newProcessorConfiguration smart constructor.

Constructors

ProcessorConfiguration'

Fields

lambda :: LambdaConfiguration

Indicates that the processor is of type Lambda.

newProcessorConfiguration

Arguments

:: LambdaConfiguration
$sel:lambda:ProcessorConfiguration'

-> ProcessorConfiguration

Create a value of ProcessorConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lambda:ProcessorConfiguration', processorConfiguration_lambda - Indicates that the processor is of type Lambda.

processorConfiguration_lambda :: Lens' ProcessorConfiguration LambdaConfiguration

Indicates that the processor is of type Lambda.

Tag

data Tag

A tag object containing a key-value pair.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Sensitive Text

The key in a tag.

value :: Sensitive Text

The value in a tag.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key in a tag.

$sel:value:Tag', tag_value - The value in a tag.

tag_key :: Lens' Tag Text

The key in a tag.

tag_value :: Lens' Tag Text

The value in a tag.