Network.AWS.ChimeSDKMessaging.Types.ChannelMessageSummary

Description

Documentation

data ChannelMessageSummary

Summary of the messages in a Channel.

See: newChannelMessageSummary smart constructor.

Constructors

ChannelMessageSummary'

Fields

status :: Maybe ChannelMessageStatusStructure

The message status. The status value is SENT for messages sent to a channel without a channel flow. For channels associated with channel flow, the value determines the processing stage.

sender :: Maybe Identity

The message sender.

content :: Maybe (Sensitive Text)

The content of the message.

redacted :: Maybe Bool

Indicates whether a message was redacted.

metadata :: Maybe (Sensitive Text)

The metadata of the message.

type' :: Maybe ChannelMessageType

The type of message.

createdTimestamp :: Maybe POSIX

The time at which the message summary was created.

messageId :: Maybe Text

The ID of the message.

lastUpdatedTimestamp :: Maybe POSIX

The time at which a message was last updated.

lastEditedTimestamp :: Maybe POSIX

The time at which a message was last edited.

newChannelMessageSummary :: ChannelMessageSummary

Create a value of ChannelMessageSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ChannelMessageSummary', channelMessageSummary_status - The message status. The status value is SENT for messages sent to a channel without a channel flow. For channels associated with channel flow, the value determines the processing stage.

$sel:sender:ChannelMessageSummary', channelMessageSummary_sender - The message sender.

$sel:content:ChannelMessageSummary', channelMessageSummary_content - The content of the message.

$sel:redacted:ChannelMessageSummary', channelMessageSummary_redacted - Indicates whether a message was redacted.

$sel:metadata:ChannelMessageSummary', channelMessageSummary_metadata - The metadata of the message.

$sel:type':ChannelMessageSummary', channelMessageSummary_type - The type of message.

$sel:createdTimestamp:ChannelMessageSummary', channelMessageSummary_createdTimestamp - The time at which the message summary was created.

$sel:messageId:ChannelMessageSummary', channelMessageSummary_messageId - The ID of the message.

$sel:lastUpdatedTimestamp:ChannelMessageSummary', channelMessageSummary_lastUpdatedTimestamp - The time at which a message was last updated.

$sel:lastEditedTimestamp:ChannelMessageSummary', channelMessageSummary_lastEditedTimestamp - The time at which a message was last edited.

channelMessageSummary_status :: Lens' ChannelMessageSummary (Maybe ChannelMessageStatusStructure)

The message status. The status value is SENT for messages sent to a channel without a channel flow. For channels associated with channel flow, the value determines the processing stage.

channelMessageSummary_sender :: Lens' ChannelMessageSummary (Maybe Identity)

The message sender.

channelMessageSummary_content :: Lens' ChannelMessageSummary (Maybe Text)

The content of the message.

channelMessageSummary_redacted :: Lens' ChannelMessageSummary (Maybe Bool)

Indicates whether a message was redacted.

channelMessageSummary_metadata :: Lens' ChannelMessageSummary (Maybe Text)

The metadata of the message.

channelMessageSummary_type :: Lens' ChannelMessageSummary (Maybe ChannelMessageType)

The type of message.

channelMessageSummary_createdTimestamp :: Lens' ChannelMessageSummary (Maybe UTCTime)

The time at which the message summary was created.

channelMessageSummary_messageId :: Lens' ChannelMessageSummary (Maybe Text)

The ID of the message.

channelMessageSummary_lastUpdatedTimestamp :: Lens' ChannelMessageSummary (Maybe UTCTime)

The time at which a message was last updated.

channelMessageSummary_lastEditedTimestamp :: Lens' ChannelMessageSummary (Maybe UTCTime)

The time at which a message was last edited.