Network.AWS.ChimeSDKMessaging.Types.ChannelSummary

Description

Documentation

data ChannelSummary

Summary of the details of a Channel.

See: newChannelSummary smart constructor.

Constructors

ChannelSummary'

Fields

mode :: Maybe ChannelMode

The mode of the channel.

channelArn :: Maybe Text

The ARN of the channel.

privacy :: Maybe ChannelPrivacy

The privacy setting of the channel.

lastMessageTimestamp :: Maybe POSIX

The time at which the last message in a channel was sent.

name :: Maybe (Sensitive Text)

The name of the channel.

metadata :: Maybe (Sensitive Text)

The metadata of the channel.

newChannelSummary :: ChannelSummary

Create a value of ChannelSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mode:ChannelSummary', channelSummary_mode - The mode of the channel.

$sel:channelArn:ChannelSummary', channelSummary_channelArn - The ARN of the channel.

$sel:privacy:ChannelSummary', channelSummary_privacy - The privacy setting of the channel.

$sel:lastMessageTimestamp:ChannelSummary', channelSummary_lastMessageTimestamp - The time at which the last message in a channel was sent.

$sel:name:ChannelSummary', channelSummary_name - The name of the channel.

$sel:metadata:ChannelSummary', channelSummary_metadata - The metadata of the channel.

channelSummary_mode :: Lens' ChannelSummary (Maybe ChannelMode)

The mode of the channel.

channelSummary_channelArn :: Lens' ChannelSummary (Maybe Text)

The ARN of the channel.

channelSummary_privacy :: Lens' ChannelSummary (Maybe ChannelPrivacy)

The privacy setting of the channel.

channelSummary_lastMessageTimestamp :: Lens' ChannelSummary (Maybe UTCTime)

The time at which the last message in a channel was sent.

channelSummary_name :: Lens' ChannelSummary (Maybe Text)

The name of the channel.

channelSummary_metadata :: Lens' ChannelSummary (Maybe Text)

The metadata of the channel.