Network.AWS.ChimeSDKMessaging.Types.Processor

Description

Documentation

data Processor

The information about a processor in a channel flow.

See: newProcessor smart constructor.

Constructors

Processor'

Fields

name :: Sensitive Text

The name of the channel flow.

configuration :: ProcessorConfiguration

The information about the type of processor and its identifier.

executionOrder :: Natural

The sequence in which processors run. If you have multiple processors in a channel flow, message processing goes through each processor in the sequence. The value determines the sequence. At this point, we support only 1 processor within a flow.

fallbackAction :: FallbackAction

Determines whether to continue or stop processing if communication with processor fails. If the last processor in a channel flow sequence has a fallback action of CONTINUE, and communication with the processor fails, the message is considered processed and sent to the recipients in the channel.

newProcessor

Arguments

:: Text
$sel:name:Processor'

-> ProcessorConfiguration
$sel:configuration:Processor'

-> Natural
$sel:executionOrder:Processor'

-> FallbackAction
$sel:fallbackAction:Processor'

-> Processor

Create a value of Processor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Processor', processor_name - The name of the channel flow.

$sel:configuration:Processor', processor_configuration - The information about the type of processor and its identifier.

$sel:executionOrder:Processor', processor_executionOrder - The sequence in which processors run. If you have multiple processors in a channel flow, message processing goes through each processor in the sequence. The value determines the sequence. At this point, we support only 1 processor within a flow.

$sel:fallbackAction:Processor', processor_fallbackAction - Determines whether to continue or stop processing if communication with processor fails. If the last processor in a channel flow sequence has a fallback action of CONTINUE, and communication with the processor fails, the message is considered processed and sent to the recipients in the channel.

processor_name :: Lens' Processor Text

The name of the channel flow.

processor_configuration :: Lens' Processor ProcessorConfiguration

The information about the type of processor and its identifier.

processor_executionOrder :: Lens' Processor Natural

The sequence in which processors run. If you have multiple processors in a channel flow, message processing goes through each processor in the sequence. The value determines the sequence. At this point, we support only 1 processor within a flow.

processor_fallbackAction :: Lens' Processor FallbackAction

Determines whether to continue or stop processing if communication with processor fails. If the last processor in a channel flow sequence has a fallback action of CONTINUE, and communication with the processor fails, the message is considered processed and sent to the recipients in the channel.