Network.AWS.Cloud9

Description

Derived from API version 2017-09-23 of the AWS service descriptions, licensed under Apache 2.0.

Cloud9

Cloud9 is a collection of tools that you can use to code, build, run, test, debug, and release software in the cloud.

For more information about Cloud9, see the Cloud9 User Guide.

Cloud9 supports these operations:

CreateEnvironmentEC2: Creates an Cloud9 development environment, launches an Amazon EC2 instance, and then connects from the instance to the environment.

CreateEnvironmentMembership: Adds an environment member to an environment.

DeleteEnvironment: Deletes an environment. If an Amazon EC2 instance is connected to the environment, also terminates the instance.

DeleteEnvironmentMembership: Deletes an environment member from an environment.

DescribeEnvironmentMemberships: Gets information about environment members for an environment.

DescribeEnvironments: Gets information about environments.

DescribeEnvironmentStatus: Gets status information for an environment.

ListEnvironments: Gets a list of environment identifiers.

ListTagsForResource: Gets the tags for an environment.

TagResource: Adds tags to an environment.

UntagResource: Removes tags from an environment.

UpdateEnvironment: Changes the settings of an existing environment.

UpdateEnvironmentMembership: Changes the settings of an existing environment member for an environment.

Service Configuration

defaultService :: Service

API version 2017-09-23 of the Amazon Cloud9 SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Cloud9.

NotFoundException

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The target resource cannot be found.

BadRequestException

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The target request is invalid.

ConcurrentAccessException

_ConcurrentAccessException :: AsError a => Getting (First ServiceError) a ServiceError

A concurrent access issue occurred.

InternalServerErrorException

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An internal server error occurred.

ForbiddenException

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

An access permissions issue occurred.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

A conflict occurred.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A service limit was exceeded.

TooManyRequestsException

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Too many service requests were made over the given time period.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListEnvironments (Paginated)

data ListEnvironments

See: newListEnvironments smart constructor.

Constructors

ListEnvironments' (Maybe Text) (Maybe Natural)

newListEnvironments :: ListEnvironments

Create a value of ListEnvironments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEnvironments', listEnvironments_nextToken - During a previous call, if there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

$sel:maxResults:ListEnvironments', listEnvironments_maxResults - The maximum number of environments to get identifiers for.

data ListEnvironmentsResponse

See: newListEnvironmentsResponse smart constructor.

Constructors

ListEnvironmentsResponse' (Maybe Text) (Maybe [Text]) Int

newListEnvironmentsResponse

Arguments

:: Int
$sel:httpStatus:ListEnvironmentsResponse'

-> ListEnvironmentsResponse

Create a value of ListEnvironmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEnvironments', listEnvironmentsResponse_nextToken - If there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call.

$sel:environmentIds:ListEnvironmentsResponse', listEnvironmentsResponse_environmentIds - The list of environment identifiers.

$sel:httpStatus:ListEnvironmentsResponse', listEnvironmentsResponse_httpStatus - The response's http status code.

CreateEnvironmentMembership

data CreateEnvironmentMembership

See: newCreateEnvironmentMembership smart constructor.

Constructors

CreateEnvironmentMembership' Text Text MemberPermissions

newCreateEnvironmentMembership

Arguments

:: Text
$sel:environmentId:CreateEnvironmentMembership'

-> Text
$sel:userArn:CreateEnvironmentMembership'

-> MemberPermissions
$sel:permissions:CreateEnvironmentMembership'

-> CreateEnvironmentMembership

Create a value of CreateEnvironmentMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:CreateEnvironmentMembership', createEnvironmentMembership_environmentId - The ID of the environment that contains the environment member you want to add.

$sel:userArn:CreateEnvironmentMembership', createEnvironmentMembership_userArn - The Amazon Resource Name (ARN) of the environment member you want to add.

$sel:permissions:CreateEnvironmentMembership', createEnvironmentMembership_permissions - The type of environment member permissions you want to associate with this environment member. Available values include:

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

data CreateEnvironmentMembershipResponse

See: newCreateEnvironmentMembershipResponse smart constructor.

Constructors

CreateEnvironmentMembershipResponse' Int EnvironmentMember

newCreateEnvironmentMembershipResponse

Arguments

:: Int
$sel:httpStatus:CreateEnvironmentMembershipResponse'

-> EnvironmentMember
$sel:membership:CreateEnvironmentMembershipResponse'

-> CreateEnvironmentMembershipResponse

Create a value of CreateEnvironmentMembershipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateEnvironmentMembershipResponse', createEnvironmentMembershipResponse_httpStatus - The response's http status code.

$sel:membership:CreateEnvironmentMembershipResponse', createEnvironmentMembershipResponse_membership - Information about the environment member that was added.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text (Sensitive [Sensitive Text])

newUntagResource

Arguments

:: Text
$sel:resourceARN:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARN:UntagResource', untagResource_resourceARN - The Amazon Resource Name (ARN) of the Cloud9 development environment to remove tags from.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The tag names of the tags to remove from the given Cloud9 development environment.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (Sensitive [Sensitive Tag])

newTagResource

Arguments

:: Text
$sel:resourceARN:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARN:TagResource', tagResource_resourceARN - The Amazon Resource Name (ARN) of the Cloud9 development environment to add tags to.

$sel:tags:TagResource', tagResource_tags - The list of tags to add to the given Cloud9 development environment.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

DeleteEnvironment

data DeleteEnvironment

See: newDeleteEnvironment smart constructor.

Constructors

DeleteEnvironment' Text

newDeleteEnvironment

Arguments

:: Text
$sel:environmentId:DeleteEnvironment'

-> DeleteEnvironment

Create a value of DeleteEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:DeleteEnvironment', deleteEnvironment_environmentId - The ID of the environment to delete.

data DeleteEnvironmentResponse

See: newDeleteEnvironmentResponse smart constructor.

Constructors

DeleteEnvironmentResponse' Int

newDeleteEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:DeleteEnvironmentResponse'

-> DeleteEnvironmentResponse

Create a value of DeleteEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteEnvironmentResponse', deleteEnvironmentResponse_httpStatus - The response's http status code.

DescribeEnvironmentStatus

data DescribeEnvironmentStatus

See: newDescribeEnvironmentStatus smart constructor.

Constructors

DescribeEnvironmentStatus' Text

newDescribeEnvironmentStatus

Arguments

:: Text
$sel:environmentId:DescribeEnvironmentStatus'

-> DescribeEnvironmentStatus

Create a value of DescribeEnvironmentStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:DescribeEnvironmentStatus', describeEnvironmentStatus_environmentId - The ID of the environment to get status information about.

data DescribeEnvironmentStatusResponse

See: newDescribeEnvironmentStatusResponse smart constructor.

Constructors

DescribeEnvironmentStatusResponse' Int EnvironmentStatus Text

newDescribeEnvironmentStatusResponse

Arguments

:: Int
$sel:httpStatus:DescribeEnvironmentStatusResponse'

-> EnvironmentStatus
$sel:status:DescribeEnvironmentStatusResponse'

-> Text
$sel:message:DescribeEnvironmentStatusResponse'

-> DescribeEnvironmentStatusResponse

Create a value of DescribeEnvironmentStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeEnvironmentStatusResponse', describeEnvironmentStatusResponse_httpStatus - The response's http status code.

$sel:status:DescribeEnvironmentStatusResponse', describeEnvironmentStatusResponse_status - The status of the environment. Available values include:

connecting: The environment is connecting.

creating: The environment is being created.

deleting: The environment is being deleted.

error: The environment is in an error state.

ready: The environment is ready.

stopped: The environment is stopped.

stopping: The environment is stopping.

$sel:message:DescribeEnvironmentStatusResponse', describeEnvironmentStatusResponse_message - Any informational message about the status of the environment.

UpdateEnvironment

data UpdateEnvironment

See: newUpdateEnvironment smart constructor.

Constructors

UpdateEnvironment' (Maybe ManagedCredentialsAction) (Maybe Text) (Maybe (Sensitive Text)) Text

newUpdateEnvironment

Arguments

:: Text
$sel:environmentId:UpdateEnvironment'

-> UpdateEnvironment

Create a value of UpdateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:managedCredentialsAction:UpdateEnvironment', updateEnvironment_managedCredentialsAction - Allows the environment owner to turn on or turn off the Amazon Web Services managed temporary credentials for an Cloud9 environment by using one of the following values:

ENABLE

DISABLE

Only the environment owner can change the status of managed temporary credentials. An AccessDeniedException is thrown if an attempt to turn on or turn off managed temporary credentials is made by an account that's not the environment owner.

$sel:name:UpdateEnvironment', updateEnvironment_name - A replacement name for the environment.

$sel:description:UpdateEnvironment', updateEnvironment_description - Any new or replacement description for the environment.

$sel:environmentId:UpdateEnvironment', updateEnvironment_environmentId - The ID of the environment to change settings.

data UpdateEnvironmentResponse

See: newUpdateEnvironmentResponse smart constructor.

Constructors

UpdateEnvironmentResponse' Int

newUpdateEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:UpdateEnvironmentResponse'

-> UpdateEnvironmentResponse

Create a value of UpdateEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateEnvironmentResponse', updateEnvironmentResponse_httpStatus - The response's http status code.

DescribeEnvironmentMemberships (Paginated)

data DescribeEnvironmentMemberships

See: newDescribeEnvironmentMemberships smart constructor.

Constructors

DescribeEnvironmentMemberships' (Maybe Text) (Maybe Text) (Maybe Natural) (Maybe Text) (Maybe [Permissions])

newDescribeEnvironmentMemberships :: DescribeEnvironmentMemberships

Create a value of DescribeEnvironmentMemberships with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userArn:DescribeEnvironmentMemberships', describeEnvironmentMemberships_userArn - The Amazon Resource Name (ARN) of an individual environment member to get information about. If no value is specified, information about all environment members are returned.

$sel:nextToken:DescribeEnvironmentMemberships', describeEnvironmentMemberships_nextToken - During a previous call, if there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

$sel:maxResults:DescribeEnvironmentMemberships', describeEnvironmentMemberships_maxResults - The maximum number of environment members to get information about.

$sel:environmentId:DescribeEnvironmentMemberships', describeEnvironmentMemberships_environmentId - The ID of the environment to get environment member information about.

$sel:permissions:DescribeEnvironmentMemberships', describeEnvironmentMemberships_permissions - The type of environment member permissions to get information about. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

If no value is specified, information about all environment members are returned.

data DescribeEnvironmentMembershipsResponse

See: newDescribeEnvironmentMembershipsResponse smart constructor.

Constructors

DescribeEnvironmentMembershipsResponse' (Maybe Text) (Maybe [EnvironmentMember]) Int

newDescribeEnvironmentMembershipsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEnvironmentMembershipsResponse'

-> DescribeEnvironmentMembershipsResponse

Create a value of DescribeEnvironmentMembershipsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeEnvironmentMemberships', describeEnvironmentMembershipsResponse_nextToken - If there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call.

$sel:memberships:DescribeEnvironmentMembershipsResponse', describeEnvironmentMembershipsResponse_memberships - Information about the environment members for the environment.

$sel:httpStatus:DescribeEnvironmentMembershipsResponse', describeEnvironmentMembershipsResponse_httpStatus - The response's http status code.

DeleteEnvironmentMembership

data DeleteEnvironmentMembership

See: newDeleteEnvironmentMembership smart constructor.

Constructors

DeleteEnvironmentMembership' Text Text

newDeleteEnvironmentMembership

Arguments

:: Text
$sel:environmentId:DeleteEnvironmentMembership'

-> Text
$sel:userArn:DeleteEnvironmentMembership'

-> DeleteEnvironmentMembership

Create a value of DeleteEnvironmentMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:DeleteEnvironmentMembership', deleteEnvironmentMembership_environmentId - The ID of the environment to delete the environment member from.

$sel:userArn:DeleteEnvironmentMembership', deleteEnvironmentMembership_userArn - The Amazon Resource Name (ARN) of the environment member to delete from the environment.

data DeleteEnvironmentMembershipResponse

See: newDeleteEnvironmentMembershipResponse smart constructor.

Constructors

DeleteEnvironmentMembershipResponse' Int

newDeleteEnvironmentMembershipResponse

Arguments

:: Int
$sel:httpStatus:DeleteEnvironmentMembershipResponse'

-> DeleteEnvironmentMembershipResponse

Create a value of DeleteEnvironmentMembershipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteEnvironmentMembershipResponse', deleteEnvironmentMembershipResponse_httpStatus - The response's http status code.

UpdateEnvironmentMembership

data UpdateEnvironmentMembership

See: newUpdateEnvironmentMembership smart constructor.

Constructors

UpdateEnvironmentMembership' Text Text MemberPermissions

newUpdateEnvironmentMembership

Arguments

:: Text
$sel:environmentId:UpdateEnvironmentMembership'

-> Text
$sel:userArn:UpdateEnvironmentMembership'

-> MemberPermissions
$sel:permissions:UpdateEnvironmentMembership'

-> UpdateEnvironmentMembership

Create a value of UpdateEnvironmentMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:UpdateEnvironmentMembership', updateEnvironmentMembership_environmentId - The ID of the environment for the environment member whose settings you want to change.

$sel:userArn:UpdateEnvironmentMembership', updateEnvironmentMembership_userArn - The Amazon Resource Name (ARN) of the environment member whose settings you want to change.

$sel:permissions:UpdateEnvironmentMembership', updateEnvironmentMembership_permissions - The replacement type of environment member permissions you want to associate with this environment member. Available values include:

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

data UpdateEnvironmentMembershipResponse

See: newUpdateEnvironmentMembershipResponse smart constructor.

Constructors

UpdateEnvironmentMembershipResponse' (Maybe EnvironmentMember) Int

newUpdateEnvironmentMembershipResponse

Arguments

:: Int
$sel:httpStatus:UpdateEnvironmentMembershipResponse'

-> UpdateEnvironmentMembershipResponse

Create a value of UpdateEnvironmentMembershipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:membership:UpdateEnvironmentMembershipResponse', updateEnvironmentMembershipResponse_membership - Information about the environment member whose settings were changed.

$sel:httpStatus:UpdateEnvironmentMembershipResponse', updateEnvironmentMembershipResponse_httpStatus - The response's http status code.

DescribeEnvironments

data DescribeEnvironments

See: newDescribeEnvironments smart constructor.

Constructors

DescribeEnvironments' (NonEmpty Text)

newDescribeEnvironments

Arguments

:: NonEmpty Text
$sel:environmentIds:DescribeEnvironments'

-> DescribeEnvironments

Create a value of DescribeEnvironments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentIds:DescribeEnvironments', describeEnvironments_environmentIds - The IDs of individual environments to get information about.

data DescribeEnvironmentsResponse

See: newDescribeEnvironmentsResponse smart constructor.

Constructors

DescribeEnvironmentsResponse' (Maybe [Environment]) Int

newDescribeEnvironmentsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEnvironmentsResponse'

-> DescribeEnvironmentsResponse

Create a value of DescribeEnvironmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environments:DescribeEnvironmentsResponse', describeEnvironmentsResponse_environments - Information about the environments that are returned.

$sel:httpStatus:DescribeEnvironmentsResponse', describeEnvironmentsResponse_httpStatus - The response's http status code.

CreateEnvironmentEC

data CreateEnvironmentEC

See: newCreateEnvironmentEC smart constructor.

Constructors

CreateEnvironmentEC' (Maybe Bool) (Maybe ConnectionType) (Maybe Text) (Maybe Text) (Maybe (Sensitive [Sensitive Tag])) (Maybe Text) (Maybe (Sensitive Text)) (Maybe Text) (Maybe Int) Text Text

newCreateEnvironmentEC

Arguments

:: Text
$sel:name:CreateEnvironmentEC'

-> Text
$sel:instanceType:CreateEnvironmentEC'

-> CreateEnvironmentEC

Create a value of CreateEnvironmentEC with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:CreateEnvironmentEC', createEnvironmentEC_dryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

$sel:connectionType:CreateEnvironmentEC', createEnvironmentEC_connectionType - The connection type used for connecting to an Amazon EC2 environment. Valid values are CONNECT_SSH (default) and CONNECT_SSM (connected through Amazon EC2 Systems Manager).

For more information, see Accessing no-ingress EC2 instances with Amazon EC2 Systems Manager in the Cloud9 User Guide.

$sel:imageId:CreateEnvironmentEC', createEnvironmentEC_imageId - The identifier for the Amazon Machine Image (AMI) that's used to create the EC2 instance. To choose an AMI for the instance, you must specify a valid AMI alias or a valid Amazon EC2 Systems Manager (SSM) path.

The default AMI is used if the parameter isn't explicitly assigned a value in the request. Because Amazon Linux AMI has ended standard support as of December 31, 2020, we recommend you choose Amazon Linux 2, which includes long term support through 2023.

AMI aliases

Amazon Linux (default): amazonlinux-1-x86_64

Amazon Linux 2: amazonlinux-2-x86_64

Ubuntu 18.04: ubuntu-18.04-x86_64

SSM paths

__Amazon Linux (default): resolve:ssm:/aws/service/cloud9/amis/amazonlinux-1-x86_64__

Amazon Linux 2: resolve:ssm:/aws/service/cloud9/amis/amazonlinux-2-x86_64

Ubuntu 18.04: resolve:ssm:/aws/service/cloud9/amis/ubuntu-18.04-x86_64

$sel:ownerArn:CreateEnvironmentEC', createEnvironmentEC_ownerArn - The Amazon Resource Name (ARN) of the environment owner. This ARN can be the ARN of any IAM principal. If this value is not specified, the ARN defaults to this environment's creator.

$sel:tags:CreateEnvironmentEC', createEnvironmentEC_tags - An array of key-value pairs that will be associated with the new Cloud9 development environment.

$sel:subnetId:CreateEnvironmentEC', createEnvironmentEC_subnetId - The ID of the subnet in Amazon VPC that Cloud9 will use to communicate with the Amazon EC2 instance.

$sel:description:CreateEnvironmentEC', createEnvironmentEC_description - The description of the environment to create.

$sel:clientRequestToken:CreateEnvironmentEC', createEnvironmentEC_clientRequestToken - A unique, case-sensitive string that helps Cloud9 to ensure this operation completes no more than one time.

For more information, see Client Tokens in the Amazon EC2 API Reference.

$sel:automaticStopTimeMinutes:CreateEnvironmentEC', createEnvironmentEC_automaticStopTimeMinutes - The number of minutes until the running instance is shut down after the environment has last been used.

$sel:name:CreateEnvironmentEC', createEnvironmentEC_name - The name of the environment to create.

This name is visible to other IAM users in the same Amazon Web Services account.

$sel:instanceType:CreateEnvironmentEC', createEnvironmentEC_instanceType - The type of instance to connect to the environment (for example, t2.micro).

data CreateEnvironmentECResponse

See: newCreateEnvironmentECResponse smart constructor.

Constructors

CreateEnvironmentECResponse' (Maybe Text) Int

newCreateEnvironmentECResponse

Arguments

:: Int
$sel:httpStatus:CreateEnvironmentECResponse'

-> CreateEnvironmentECResponse

Create a value of CreateEnvironmentECResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:CreateEnvironmentECResponse', createEnvironmentECResponse_environmentId - The ID of the environment that was created.

$sel:httpStatus:CreateEnvironmentECResponse', createEnvironmentECResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceARN:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARN:ListTagsForResource', listTagsForResource_resourceARN - The Amazon Resource Name (ARN) of the Cloud9 development environment to get the tags for.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (Sensitive [Sensitive Tag])) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The list of tags associated with the Cloud9 development environment.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

Types

ConnectionType

newtype ConnectionType

Constructors

ConnectionType'

Fields

fromConnectionType :: Text

Bundled Patterns

pattern ConnectionType_CONNECT_SSH :: ConnectionType

pattern ConnectionType_CONNECT_SSM :: ConnectionType

EnvironmentLifecycleStatus

newtype EnvironmentLifecycleStatus

Constructors

EnvironmentLifecycleStatus'

Fields

fromEnvironmentLifecycleStatus :: Text

Bundled Patterns

pattern EnvironmentLifecycleStatus_CREATED :: EnvironmentLifecycleStatus

pattern EnvironmentLifecycleStatus_CREATE_FAILED :: EnvironmentLifecycleStatus

pattern EnvironmentLifecycleStatus_CREATING :: EnvironmentLifecycleStatus

pattern EnvironmentLifecycleStatus_DELETE_FAILED :: EnvironmentLifecycleStatus

pattern EnvironmentLifecycleStatus_DELETING :: EnvironmentLifecycleStatus

EnvironmentStatus

newtype EnvironmentStatus

Constructors

EnvironmentStatus'

Fields

fromEnvironmentStatus :: Text

Bundled Patterns

pattern EnvironmentStatus_Connecting :: EnvironmentStatus

pattern EnvironmentStatus_Creating :: EnvironmentStatus

pattern EnvironmentStatus_Deleting :: EnvironmentStatus

pattern EnvironmentStatus_Error :: EnvironmentStatus

pattern EnvironmentStatus_Ready :: EnvironmentStatus

pattern EnvironmentStatus_Stopped :: EnvironmentStatus

pattern EnvironmentStatus_Stopping :: EnvironmentStatus

EnvironmentType

newtype EnvironmentType

Constructors

EnvironmentType'

Fields

fromEnvironmentType :: Text

Bundled Patterns

pattern EnvironmentType_Ec2 :: EnvironmentType

pattern EnvironmentType_Ssh :: EnvironmentType

ManagedCredentialsAction

newtype ManagedCredentialsAction

Constructors

ManagedCredentialsAction'

Fields

fromManagedCredentialsAction :: Text

Bundled Patterns

pattern ManagedCredentialsAction_DISABLE :: ManagedCredentialsAction

pattern ManagedCredentialsAction_ENABLE :: ManagedCredentialsAction

ManagedCredentialsStatus

newtype ManagedCredentialsStatus

Constructors

ManagedCredentialsStatus'

Fields

fromManagedCredentialsStatus :: Text

Bundled Patterns

pattern ManagedCredentialsStatus_DISABLED_BY_COLLABORATOR :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_DISABLED_BY_DEFAULT :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_DISABLED_BY_OWNER :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_ENABLED_BY_OWNER :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_ENABLED_ON_CREATE :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_FAILED_REMOVAL_BY_COLLABORATOR :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_FAILED_REMOVAL_BY_OWNER :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_PENDING_REMOVAL_BY_COLLABORATOR :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_PENDING_REMOVAL_BY_OWNER :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_PENDING_START_REMOVAL_BY_COLLABORATOR :: ManagedCredentialsStatus

pattern ManagedCredentialsStatus_PENDING_START_REMOVAL_BY_OWNER :: ManagedCredentialsStatus

MemberPermissions

newtype MemberPermissions

Constructors

MemberPermissions'

Fields

fromMemberPermissions :: Text

Bundled Patterns

pattern MemberPermissions_Read_only :: MemberPermissions

pattern MemberPermissions_Read_write :: MemberPermissions

Permissions

newtype Permissions

Constructors

Permissions'

Fields

fromPermissions :: Text

Bundled Patterns

pattern Permissions_Owner :: Permissions

pattern Permissions_Read_only :: Permissions

pattern Permissions_Read_write :: Permissions

Environment

data Environment

Information about an Cloud9 development environment.

See: newEnvironment smart constructor.

Constructors

Environment' (Maybe EnvironmentLifecycle) (Maybe ConnectionType) (Maybe Text) (Maybe Text) (Maybe ManagedCredentialsStatus) (Maybe (Sensitive Text)) EnvironmentType Text Text

newEnvironment

Arguments

:: EnvironmentType
$sel:type':Environment'

-> Text
$sel:arn:Environment'

-> Text
$sel:ownerArn:Environment'

-> Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lifecycle:Environment', environment_lifecycle - The state of the environment in its creation or deletion lifecycle.

$sel:connectionType:Environment', environment_connectionType - The connection type used for connecting to an Amazon EC2 environment. CONNECT_SSH is selected by default.

$sel:id:Environment', environment_id - The ID of the environment.

$sel:name:Environment', environment_name - The name of the environment.

$sel:managedCredentialsStatus:Environment', environment_managedCredentialsStatus - Describes the status of Amazon Web Services managed temporary credentials for the Cloud9 environment. Available values are:

ENABLED_ON_CREATE

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

DISABLED_BY_OWNER

DISABLED_BY_COLLABORATOR

PENDING_REMOVAL_BY_COLLABORATOR

PENDING_REMOVAL_BY_OWNER

FAILED_REMOVAL_BY_COLLABORATOR

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

$sel:description:Environment', environment_description - The description for the environment.

$sel:type':Environment', environment_type - The type of environment. Valid values include the following:

ec2: An Amazon Elastic Compute Cloud (Amazon EC2) instance connects to the environment.

ssh: Your own server connects to the environment.

$sel:arn:Environment', environment_arn - The Amazon Resource Name (ARN) of the environment.

$sel:ownerArn:Environment', environment_ownerArn - The Amazon Resource Name (ARN) of the environment owner.

EnvironmentLifecycle

data EnvironmentLifecycle

Information about the current creation or deletion lifecycle state of an Cloud9 development environment.

See: newEnvironmentLifecycle smart constructor.

Constructors

EnvironmentLifecycle' (Maybe EnvironmentLifecycleStatus) (Maybe Text) (Maybe Text)

newEnvironmentLifecycle :: EnvironmentLifecycle

Create a value of EnvironmentLifecycle with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:EnvironmentLifecycle', environmentLifecycle_status - The current creation or deletion lifecycle state of the environment.

CREATING: The environment is in the process of being created.

CREATED: The environment was successfully created.

CREATE_FAILED: The environment failed to be created.

DELETING: The environment is in the process of being deleted.

DELETE_FAILED: The environment failed to delete.

$sel:reason:EnvironmentLifecycle', environmentLifecycle_reason - Any informational message about the lifecycle state of the environment.

$sel:failureResource:EnvironmentLifecycle', environmentLifecycle_failureResource - If the environment failed to delete, the Amazon Resource Name (ARN) of the related Amazon Web Services resource.

EnvironmentMember

data EnvironmentMember

Information about an environment member for an Cloud9 development environment.

See: newEnvironmentMember smart constructor.

Constructors

EnvironmentMember' (Maybe POSIX) Permissions Text Text Text

newEnvironmentMember

Arguments

:: Permissions
$sel:permissions:EnvironmentMember'

-> Text
$sel:userId:EnvironmentMember'

-> Text
$sel:userArn:EnvironmentMember'

-> Text
$sel:environmentId:EnvironmentMember'

-> EnvironmentMember

Create a value of EnvironmentMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastAccess:EnvironmentMember', environmentMember_lastAccess - The time, expressed in epoch time format, when the environment member last opened the environment.

$sel:permissions:EnvironmentMember', environmentMember_permissions - The type of environment member permissions associated with this environment member. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

$sel:userId:EnvironmentMember', environmentMember_userId - The user ID in Identity and Access Management (IAM) of the environment member.

$sel:userArn:EnvironmentMember', environmentMember_userArn - The Amazon Resource Name (ARN) of the environment member.

$sel:environmentId:EnvironmentMember', environmentMember_environmentId - The ID of the environment for the environment member.

Tag

data Tag

Metadata that is associated with Amazon Web Services resources. In particular, a name-value pair that can be associated with an Cloud9 development environment. There are two types of tags: user tags and system tags. A user tag is created by the user. A system tag is automatically created by Amazon Web Services services. A system tag is prefixed with "aws:" and cannot be modified by the user.

See: newTag smart constructor.

Constructors

Tag' (Sensitive Text) (Sensitive Text)

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The name part of a tag.

$sel:value:Tag', tag_value - The value part of a tag.