Network.AWS.Cloud9.CreateEnvironmentMembership

Description

Adds an environment member to an Cloud9 development environment.

Creating a Request

data CreateEnvironmentMembership

See: newCreateEnvironmentMembership smart constructor.

Constructors

CreateEnvironmentMembership'

Fields

environmentId :: Text

The ID of the environment that contains the environment member you want to add.

userArn :: Text

The Amazon Resource Name (ARN) of the environment member you want to add.

permissions :: MemberPermissions

The type of environment member permissions you want to associate with this environment member. Available values include:

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

newCreateEnvironmentMembership

Arguments

:: Text
$sel:environmentId:CreateEnvironmentMembership'

-> Text
$sel:userArn:CreateEnvironmentMembership'

-> MemberPermissions
$sel:permissions:CreateEnvironmentMembership'

-> CreateEnvironmentMembership

Create a value of CreateEnvironmentMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:CreateEnvironmentMembership', createEnvironmentMembership_environmentId - The ID of the environment that contains the environment member you want to add.

$sel:userArn:CreateEnvironmentMembership', createEnvironmentMembership_userArn - The Amazon Resource Name (ARN) of the environment member you want to add.

$sel:permissions:CreateEnvironmentMembership', createEnvironmentMembership_permissions - The type of environment member permissions you want to associate with this environment member. Available values include:

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

Request Lenses

createEnvironmentMembership_environmentId :: Lens' CreateEnvironmentMembership Text

The ID of the environment that contains the environment member you want to add.

createEnvironmentMembership_userArn :: Lens' CreateEnvironmentMembership Text

The Amazon Resource Name (ARN) of the environment member you want to add.

createEnvironmentMembership_permissions :: Lens' CreateEnvironmentMembership MemberPermissions

The type of environment member permissions you want to associate with this environment member. Available values include:

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

Destructuring the Response

data CreateEnvironmentMembershipResponse

See: newCreateEnvironmentMembershipResponse smart constructor.

Constructors

CreateEnvironmentMembershipResponse'

Fields

httpStatus :: Int

The response's http status code.

membership :: EnvironmentMember

Information about the environment member that was added.

newCreateEnvironmentMembershipResponse

Arguments

:: Int
$sel:httpStatus:CreateEnvironmentMembershipResponse'

-> EnvironmentMember
$sel:membership:CreateEnvironmentMembershipResponse'

-> CreateEnvironmentMembershipResponse

Create a value of CreateEnvironmentMembershipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateEnvironmentMembershipResponse', createEnvironmentMembershipResponse_httpStatus - The response's http status code.

$sel:membership:CreateEnvironmentMembershipResponse', createEnvironmentMembershipResponse_membership - Information about the environment member that was added.

Response Lenses

createEnvironmentMembershipResponse_httpStatus :: Lens' CreateEnvironmentMembershipResponse Int

The response's http status code.

createEnvironmentMembershipResponse_membership :: Lens' CreateEnvironmentMembershipResponse EnvironmentMember

Information about the environment member that was added.