Network.AWS.Cloud9.DescribeEnvironmentMemberships

Description

Gets information about environment members for an Cloud9 development environment.

This operation returns paginated results.

Creating a Request

data DescribeEnvironmentMemberships

See: newDescribeEnvironmentMemberships smart constructor.

Constructors

DescribeEnvironmentMemberships'

Fields

userArn :: Maybe Text

The Amazon Resource Name (ARN) of an individual environment member to get information about. If no value is specified, information about all environment members are returned.

nextToken :: Maybe Text

During a previous call, if there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

maxResults :: Maybe Natural

The maximum number of environment members to get information about.

environmentId :: Maybe Text

The ID of the environment to get environment member information about.

permissions :: Maybe [Permissions]

The type of environment member permissions to get information about. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

If no value is specified, information about all environment members are returned.

newDescribeEnvironmentMemberships :: DescribeEnvironmentMemberships

Create a value of DescribeEnvironmentMemberships with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userArn:DescribeEnvironmentMemberships', describeEnvironmentMemberships_userArn - The Amazon Resource Name (ARN) of an individual environment member to get information about. If no value is specified, information about all environment members are returned.

$sel:nextToken:DescribeEnvironmentMemberships', describeEnvironmentMemberships_nextToken - During a previous call, if there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

$sel:maxResults:DescribeEnvironmentMemberships', describeEnvironmentMemberships_maxResults - The maximum number of environment members to get information about.

$sel:environmentId:DescribeEnvironmentMemberships', describeEnvironmentMemberships_environmentId - The ID of the environment to get environment member information about.

$sel:permissions:DescribeEnvironmentMemberships', describeEnvironmentMemberships_permissions - The type of environment member permissions to get information about. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

If no value is specified, information about all environment members are returned.

Request Lenses

describeEnvironmentMemberships_userArn :: Lens' DescribeEnvironmentMemberships (Maybe Text)

The Amazon Resource Name (ARN) of an individual environment member to get information about. If no value is specified, information about all environment members are returned.

describeEnvironmentMemberships_nextToken :: Lens' DescribeEnvironmentMemberships (Maybe Text)

During a previous call, if there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

describeEnvironmentMemberships_maxResults :: Lens' DescribeEnvironmentMemberships (Maybe Natural)

The maximum number of environment members to get information about.

describeEnvironmentMemberships_environmentId :: Lens' DescribeEnvironmentMemberships (Maybe Text)

The ID of the environment to get environment member information about.

describeEnvironmentMemberships_permissions :: Lens' DescribeEnvironmentMemberships (Maybe [Permissions])

The type of environment member permissions to get information about. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

If no value is specified, information about all environment members are returned.

Destructuring the Response

data DescribeEnvironmentMembershipsResponse

See: newDescribeEnvironmentMembershipsResponse smart constructor.

Constructors

DescribeEnvironmentMembershipsResponse'

Fields

nextToken :: Maybe Text

If there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call.

memberships :: Maybe [EnvironmentMember]

Information about the environment members for the environment.

httpStatus :: Int

The response's http status code.

newDescribeEnvironmentMembershipsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEnvironmentMembershipsResponse'

-> DescribeEnvironmentMembershipsResponse

Create a value of DescribeEnvironmentMembershipsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeEnvironmentMemberships', describeEnvironmentMembershipsResponse_nextToken - If there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call.

$sel:memberships:DescribeEnvironmentMembershipsResponse', describeEnvironmentMembershipsResponse_memberships - Information about the environment members for the environment.

$sel:httpStatus:DescribeEnvironmentMembershipsResponse', describeEnvironmentMembershipsResponse_httpStatus - The response's http status code.

Response Lenses

describeEnvironmentMembershipsResponse_nextToken :: Lens' DescribeEnvironmentMembershipsResponse (Maybe Text)

If there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call.

describeEnvironmentMembershipsResponse_memberships :: Lens' DescribeEnvironmentMembershipsResponse (Maybe [EnvironmentMember])

Information about the environment members for the environment.

describeEnvironmentMembershipsResponse_httpStatus :: Lens' DescribeEnvironmentMembershipsResponse Int

The response's http status code.