Network.AWS.Cloud9.DescribeEnvironmentStatus

Description

Gets status information for an Cloud9 development environment.

Creating a Request

data DescribeEnvironmentStatus

See: newDescribeEnvironmentStatus smart constructor.

Constructors

DescribeEnvironmentStatus'

Fields

environmentId :: Text

The ID of the environment to get status information about.

newDescribeEnvironmentStatus

Arguments

:: Text
$sel:environmentId:DescribeEnvironmentStatus'

-> DescribeEnvironmentStatus

Create a value of DescribeEnvironmentStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:environmentId:DescribeEnvironmentStatus', describeEnvironmentStatus_environmentId - The ID of the environment to get status information about.

Request Lenses

describeEnvironmentStatus_environmentId :: Lens' DescribeEnvironmentStatus Text

The ID of the environment to get status information about.

Destructuring the Response

data DescribeEnvironmentStatusResponse

See: newDescribeEnvironmentStatusResponse smart constructor.

Constructors

DescribeEnvironmentStatusResponse'

Fields

httpStatus :: Int

The response's http status code.

status :: EnvironmentStatus

The status of the environment. Available values include:

connecting: The environment is connecting.

creating: The environment is being created.

deleting: The environment is being deleted.

error: The environment is in an error state.

ready: The environment is ready.

stopped: The environment is stopped.

stopping: The environment is stopping.

message :: Text

Any informational message about the status of the environment.

newDescribeEnvironmentStatusResponse

Arguments

:: Int
$sel:httpStatus:DescribeEnvironmentStatusResponse'

-> EnvironmentStatus
$sel:status:DescribeEnvironmentStatusResponse'

-> Text
$sel:message:DescribeEnvironmentStatusResponse'

-> DescribeEnvironmentStatusResponse

Create a value of DescribeEnvironmentStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeEnvironmentStatusResponse', describeEnvironmentStatusResponse_httpStatus - The response's http status code.

$sel:status:DescribeEnvironmentStatusResponse', describeEnvironmentStatusResponse_status - The status of the environment. Available values include:

connecting: The environment is connecting.

creating: The environment is being created.

deleting: The environment is being deleted.

error: The environment is in an error state.

ready: The environment is ready.

stopped: The environment is stopped.

stopping: The environment is stopping.

$sel:message:DescribeEnvironmentStatusResponse', describeEnvironmentStatusResponse_message - Any informational message about the status of the environment.

Response Lenses

describeEnvironmentStatusResponse_httpStatus :: Lens' DescribeEnvironmentStatusResponse Int

The response's http status code.

describeEnvironmentStatusResponse_status :: Lens' DescribeEnvironmentStatusResponse EnvironmentStatus

The status of the environment. Available values include:

connecting: The environment is connecting.

creating: The environment is being created.

deleting: The environment is being deleted.

error: The environment is in an error state.

ready: The environment is ready.

stopped: The environment is stopped.

stopping: The environment is stopping.

describeEnvironmentStatusResponse_message :: Lens' DescribeEnvironmentStatusResponse Text

Any informational message about the status of the environment.