Network.AWS.Cloud9.Lens

Description

Operations

ListEnvironments

listEnvironments_nextToken :: Lens' ListEnvironments (Maybe Text)

During a previous call, if there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

listEnvironments_maxResults :: Lens' ListEnvironments (Maybe Natural)

The maximum number of environments to get identifiers for.

listEnvironmentsResponse_environmentIds :: Lens' ListEnvironmentsResponse (Maybe [Text])

The list of environment identifiers.

listEnvironmentsResponse_nextToken :: Lens' ListEnvironmentsResponse (Maybe Text)

If there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call.

listEnvironmentsResponse_httpStatus :: Lens' ListEnvironmentsResponse Int

The response's http status code.

UpdateEnvironment

updateEnvironment_managedCredentialsAction :: Lens' UpdateEnvironment (Maybe ManagedCredentialsAction)

Allows the environment owner to turn on or turn off the Amazon Web Services managed temporary credentials for an Cloud9 environment by using one of the following values:

ENABLE

DISABLE

Only the environment owner can change the status of managed temporary credentials. An AccessDeniedException is thrown if an attempt to turn on or turn off managed temporary credentials is made by an account that's not the environment owner.

updateEnvironment_name :: Lens' UpdateEnvironment (Maybe Text)

A replacement name for the environment.

updateEnvironment_description :: Lens' UpdateEnvironment (Maybe Text)

Any new or replacement description for the environment.

updateEnvironment_environmentId :: Lens' UpdateEnvironment Text

The ID of the environment to change settings.

updateEnvironmentResponse_httpStatus :: Lens' UpdateEnvironmentResponse Int

The response's http status code.

DeleteEnvironment

deleteEnvironment_environmentId :: Lens' DeleteEnvironment Text

The ID of the environment to delete.

deleteEnvironmentResponse_httpStatus :: Lens' DeleteEnvironmentResponse Int

The response's http status code.

DescribeEnvironmentStatus

describeEnvironmentStatus_environmentId :: Lens' DescribeEnvironmentStatus Text

The ID of the environment to get status information about.

describeEnvironmentStatusResponse_httpStatus :: Lens' DescribeEnvironmentStatusResponse Int

The response's http status code.

describeEnvironmentStatusResponse_status :: Lens' DescribeEnvironmentStatusResponse EnvironmentStatus

The status of the environment. Available values include:

connecting: The environment is connecting.

creating: The environment is being created.

deleting: The environment is being deleted.

error: The environment is in an error state.

ready: The environment is ready.

stopped: The environment is stopped.

stopping: The environment is stopping.

describeEnvironmentStatusResponse_message :: Lens' DescribeEnvironmentStatusResponse Text

Any informational message about the status of the environment.

ListTagsForResource

listTagsForResource_resourceARN :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the Cloud9 development environment to get the tags for.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

The list of tags associated with the Cloud9 development environment.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CreateEnvironmentEC

createEnvironmentEC_automaticStopTimeMinutes :: Lens' CreateEnvironmentEC (Maybe Int)

The number of minutes until the running instance is shut down after the environment has last been used.

createEnvironmentEC_subnetId :: Lens' CreateEnvironmentEC (Maybe Text)

The ID of the subnet in Amazon VPC that Cloud9 will use to communicate with the Amazon EC2 instance.

createEnvironmentEC_ownerArn :: Lens' CreateEnvironmentEC (Maybe Text)

The Amazon Resource Name (ARN) of the environment owner. This ARN can be the ARN of any IAM principal. If this value is not specified, the ARN defaults to this environment's creator.

createEnvironmentEC_imageId :: Lens' CreateEnvironmentEC (Maybe Text)

The identifier for the Amazon Machine Image (AMI) that's used to create the EC2 instance. To choose an AMI for the instance, you must specify a valid AMI alias or a valid Amazon EC2 Systems Manager (SSM) path.

The default AMI is used if the parameter isn't explicitly assigned a value in the request. Because Amazon Linux AMI has ended standard support as of December 31, 2020, we recommend you choose Amazon Linux 2, which includes long term support through 2023.

AMI aliases

Amazon Linux (default): amazonlinux-1-x86_64

Amazon Linux 2: amazonlinux-2-x86_64

Ubuntu 18.04: ubuntu-18.04-x86_64

SSM paths

__Amazon Linux (default): resolve:ssm:/aws/service/cloud9/amis/amazonlinux-1-x86_64__

Amazon Linux 2: resolve:ssm:/aws/service/cloud9/amis/amazonlinux-2-x86_64

Ubuntu 18.04: resolve:ssm:/aws/service/cloud9/amis/ubuntu-18.04-x86_64

createEnvironmentEC_clientRequestToken :: Lens' CreateEnvironmentEC (Maybe Text)

A unique, case-sensitive string that helps Cloud9 to ensure this operation completes no more than one time.

For more information, see Client Tokens in the Amazon EC2 API Reference.

createEnvironmentEC_connectionType :: Lens' CreateEnvironmentEC (Maybe ConnectionType)

The connection type used for connecting to an Amazon EC2 environment. Valid values are CONNECT_SSH (default) and CONNECT_SSM (connected through Amazon EC2 Systems Manager).

For more information, see Accessing no-ingress EC2 instances with Amazon EC2 Systems Manager in the Cloud9 User Guide.

createEnvironmentEC_description :: Lens' CreateEnvironmentEC (Maybe Text)

The description of the environment to create.

createEnvironmentEC_dryRun :: Lens' CreateEnvironmentEC (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

createEnvironmentEC_tags :: Lens' CreateEnvironmentEC (Maybe [Tag])

An array of key-value pairs that will be associated with the new Cloud9 development environment.

createEnvironmentEC_name :: Lens' CreateEnvironmentEC Text

The name of the environment to create.

This name is visible to other IAM users in the same Amazon Web Services account.

createEnvironmentEC_instanceType :: Lens' CreateEnvironmentEC Text

The type of instance to connect to the environment (for example, t2.micro).

createEnvironmentECResponse_environmentId :: Lens' CreateEnvironmentECResponse (Maybe Text)

The ID of the environment that was created.

createEnvironmentECResponse_httpStatus :: Lens' CreateEnvironmentECResponse Int

The response's http status code.

TagResource

tagResource_resourceARN :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the Cloud9 development environment to add tags to.

tagResource_tags :: Lens' TagResource [Tag]

The list of tags to add to the given Cloud9 development environment.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

CreateEnvironmentMembership

createEnvironmentMembership_environmentId :: Lens' CreateEnvironmentMembership Text

The ID of the environment that contains the environment member you want to add.

createEnvironmentMembership_userArn :: Lens' CreateEnvironmentMembership Text

The Amazon Resource Name (ARN) of the environment member you want to add.

createEnvironmentMembership_permissions :: Lens' CreateEnvironmentMembership MemberPermissions

The type of environment member permissions you want to associate with this environment member. Available values include:

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

createEnvironmentMembershipResponse_httpStatus :: Lens' CreateEnvironmentMembershipResponse Int

The response's http status code.

createEnvironmentMembershipResponse_membership :: Lens' CreateEnvironmentMembershipResponse EnvironmentMember

Information about the environment member that was added.

DescribeEnvironments

describeEnvironments_environmentIds :: Lens' DescribeEnvironments (NonEmpty Text)

The IDs of individual environments to get information about.

describeEnvironmentsResponse_environments :: Lens' DescribeEnvironmentsResponse (Maybe [Environment])

Information about the environments that are returned.

describeEnvironmentsResponse_httpStatus :: Lens' DescribeEnvironmentsResponse Int

The response's http status code.

UntagResource

untagResource_resourceARN :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the Cloud9 development environment to remove tags from.

untagResource_tagKeys :: Lens' UntagResource [Text]

The tag names of the tags to remove from the given Cloud9 development environment.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DeleteEnvironmentMembership

deleteEnvironmentMembership_environmentId :: Lens' DeleteEnvironmentMembership Text

The ID of the environment to delete the environment member from.

deleteEnvironmentMembership_userArn :: Lens' DeleteEnvironmentMembership Text

The Amazon Resource Name (ARN) of the environment member to delete from the environment.

deleteEnvironmentMembershipResponse_httpStatus :: Lens' DeleteEnvironmentMembershipResponse Int

The response's http status code.

UpdateEnvironmentMembership

updateEnvironmentMembership_environmentId :: Lens' UpdateEnvironmentMembership Text

The ID of the environment for the environment member whose settings you want to change.

updateEnvironmentMembership_userArn :: Lens' UpdateEnvironmentMembership Text

The Amazon Resource Name (ARN) of the environment member whose settings you want to change.

updateEnvironmentMembership_permissions :: Lens' UpdateEnvironmentMembership MemberPermissions

The replacement type of environment member permissions you want to associate with this environment member. Available values include:

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

updateEnvironmentMembershipResponse_membership :: Lens' UpdateEnvironmentMembershipResponse (Maybe EnvironmentMember)

Information about the environment member whose settings were changed.

updateEnvironmentMembershipResponse_httpStatus :: Lens' UpdateEnvironmentMembershipResponse Int

The response's http status code.

DescribeEnvironmentMemberships

describeEnvironmentMemberships_userArn :: Lens' DescribeEnvironmentMemberships (Maybe Text)

The Amazon Resource Name (ARN) of an individual environment member to get information about. If no value is specified, information about all environment members are returned.

describeEnvironmentMemberships_nextToken :: Lens' DescribeEnvironmentMemberships (Maybe Text)

During a previous call, if there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

describeEnvironmentMemberships_permissions :: Lens' DescribeEnvironmentMemberships (Maybe [Permissions])

The type of environment member permissions to get information about. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

If no value is specified, information about all environment members are returned.

describeEnvironmentMemberships_environmentId :: Lens' DescribeEnvironmentMemberships (Maybe Text)

The ID of the environment to get environment member information about.

describeEnvironmentMemberships_maxResults :: Lens' DescribeEnvironmentMemberships (Maybe Natural)

The maximum number of environment members to get information about.

describeEnvironmentMembershipsResponse_nextToken :: Lens' DescribeEnvironmentMembershipsResponse (Maybe Text)

If there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call.

describeEnvironmentMembershipsResponse_memberships :: Lens' DescribeEnvironmentMembershipsResponse (Maybe [EnvironmentMember])

Information about the environment members for the environment.

describeEnvironmentMembershipsResponse_httpStatus :: Lens' DescribeEnvironmentMembershipsResponse Int

The response's http status code.

Types

Environment

environment_lifecycle :: Lens' Environment (Maybe EnvironmentLifecycle)

The state of the environment in its creation or deletion lifecycle.

environment_name :: Lens' Environment (Maybe Text)

The name of the environment.

environment_id :: Lens' Environment (Maybe Text)

The ID of the environment.

environment_connectionType :: Lens' Environment (Maybe ConnectionType)

The connection type used for connecting to an Amazon EC2 environment. CONNECT_SSH is selected by default.

environment_description :: Lens' Environment (Maybe Text)

The description for the environment.

environment_managedCredentialsStatus :: Lens' Environment (Maybe ManagedCredentialsStatus)

Describes the status of Amazon Web Services managed temporary credentials for the Cloud9 environment. Available values are:

ENABLED_ON_CREATE

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

DISABLED_BY_OWNER

DISABLED_BY_COLLABORATOR

PENDING_REMOVAL_BY_COLLABORATOR

PENDING_REMOVAL_BY_OWNER

FAILED_REMOVAL_BY_COLLABORATOR

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

environment_type :: Lens' Environment EnvironmentType

The type of environment. Valid values include the following:

ec2: An Amazon Elastic Compute Cloud (Amazon EC2) instance connects to the environment.

ssh: Your own server connects to the environment.

environment_arn :: Lens' Environment Text

The Amazon Resource Name (ARN) of the environment.

environment_ownerArn :: Lens' Environment Text

The Amazon Resource Name (ARN) of the environment owner.

EnvironmentLifecycle

environmentLifecycle_status :: Lens' EnvironmentLifecycle (Maybe EnvironmentLifecycleStatus)

The current creation or deletion lifecycle state of the environment.

CREATING: The environment is in the process of being created.

CREATED: The environment was successfully created.

CREATE_FAILED: The environment failed to be created.

DELETING: The environment is in the process of being deleted.

DELETE_FAILED: The environment failed to delete.

environmentLifecycle_failureResource :: Lens' EnvironmentLifecycle (Maybe Text)

If the environment failed to delete, the Amazon Resource Name (ARN) of the related Amazon Web Services resource.

environmentLifecycle_reason :: Lens' EnvironmentLifecycle (Maybe Text)

Any informational message about the lifecycle state of the environment.

EnvironmentMember

environmentMember_lastAccess :: Lens' EnvironmentMember (Maybe UTCTime)

The time, expressed in epoch time format, when the environment member last opened the environment.

environmentMember_permissions :: Lens' EnvironmentMember Permissions

The type of environment member permissions associated with this environment member. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

environmentMember_userId :: Lens' EnvironmentMember Text

The user ID in Identity and Access Management (IAM) of the environment member.

environmentMember_userArn :: Lens' EnvironmentMember Text

The Amazon Resource Name (ARN) of the environment member.

environmentMember_environmentId :: Lens' EnvironmentMember Text

The ID of the environment for the environment member.

Tag

tag_key :: Lens' Tag Text

The name part of a tag.

tag_value :: Lens' Tag Text

The value part of a tag.