Network.AWS.Cloud9.Types.Environment

Description

Documentation

data Environment

Information about an Cloud9 development environment.

See: newEnvironment smart constructor.

Constructors

Environment'

Fields

lifecycle :: Maybe EnvironmentLifecycle

The state of the environment in its creation or deletion lifecycle.

connectionType :: Maybe ConnectionType

The connection type used for connecting to an Amazon EC2 environment. CONNECT_SSH is selected by default.

id :: Maybe Text

The ID of the environment.

name :: Maybe Text

The name of the environment.

managedCredentialsStatus :: Maybe ManagedCredentialsStatus

Describes the status of Amazon Web Services managed temporary credentials for the Cloud9 environment. Available values are:

ENABLED_ON_CREATE

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

DISABLED_BY_OWNER

DISABLED_BY_COLLABORATOR

PENDING_REMOVAL_BY_COLLABORATOR

PENDING_REMOVAL_BY_OWNER

FAILED_REMOVAL_BY_COLLABORATOR

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

description :: Maybe (Sensitive Text)

The description for the environment.

type' :: EnvironmentType

The type of environment. Valid values include the following:

ec2: An Amazon Elastic Compute Cloud (Amazon EC2) instance connects to the environment.

ssh: Your own server connects to the environment.

arn :: Text

The Amazon Resource Name (ARN) of the environment.

ownerArn :: Text

The Amazon Resource Name (ARN) of the environment owner.

newEnvironment

Arguments

:: EnvironmentType
$sel:type':Environment'

-> Text
$sel:arn:Environment'

-> Text
$sel:ownerArn:Environment'

-> Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lifecycle:Environment', environment_lifecycle - The state of the environment in its creation or deletion lifecycle.

$sel:connectionType:Environment', environment_connectionType - The connection type used for connecting to an Amazon EC2 environment. CONNECT_SSH is selected by default.

$sel:id:Environment', environment_id - The ID of the environment.

$sel:name:Environment', environment_name - The name of the environment.

$sel:managedCredentialsStatus:Environment', environment_managedCredentialsStatus - Describes the status of Amazon Web Services managed temporary credentials for the Cloud9 environment. Available values are:

ENABLED_ON_CREATE

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

DISABLED_BY_OWNER

DISABLED_BY_COLLABORATOR

PENDING_REMOVAL_BY_COLLABORATOR

PENDING_REMOVAL_BY_OWNER

FAILED_REMOVAL_BY_COLLABORATOR

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

$sel:description:Environment', environment_description - The description for the environment.

$sel:type':Environment', environment_type - The type of environment. Valid values include the following:

ec2: An Amazon Elastic Compute Cloud (Amazon EC2) instance connects to the environment.

ssh: Your own server connects to the environment.

$sel:arn:Environment', environment_arn - The Amazon Resource Name (ARN) of the environment.

$sel:ownerArn:Environment', environment_ownerArn - The Amazon Resource Name (ARN) of the environment owner.

environment_lifecycle :: Lens' Environment (Maybe EnvironmentLifecycle)

The state of the environment in its creation or deletion lifecycle.

environment_connectionType :: Lens' Environment (Maybe ConnectionType)

The connection type used for connecting to an Amazon EC2 environment. CONNECT_SSH is selected by default.

environment_id :: Lens' Environment (Maybe Text)

The ID of the environment.

environment_name :: Lens' Environment (Maybe Text)

The name of the environment.

environment_managedCredentialsStatus :: Lens' Environment (Maybe ManagedCredentialsStatus)

Describes the status of Amazon Web Services managed temporary credentials for the Cloud9 environment. Available values are:

ENABLED_ON_CREATE

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

DISABLED_BY_OWNER

DISABLED_BY_COLLABORATOR

PENDING_REMOVAL_BY_COLLABORATOR

PENDING_REMOVAL_BY_OWNER

FAILED_REMOVAL_BY_COLLABORATOR

ENABLED_BY_OWNER

DISABLED_BY_DEFAULT

environment_description :: Lens' Environment (Maybe Text)

The description for the environment.

environment_type :: Lens' Environment EnvironmentType

The type of environment. Valid values include the following:

ec2: An Amazon Elastic Compute Cloud (Amazon EC2) instance connects to the environment.

ssh: Your own server connects to the environment.

environment_arn :: Lens' Environment Text

The Amazon Resource Name (ARN) of the environment.

environment_ownerArn :: Lens' Environment Text

The Amazon Resource Name (ARN) of the environment owner.