Network.AWS.Cloud9.Types.EnvironmentMember

Description

Documentation

data EnvironmentMember

Information about an environment member for an Cloud9 development environment.

See: newEnvironmentMember smart constructor.

Constructors

EnvironmentMember'

Fields

lastAccess :: Maybe POSIX

The time, expressed in epoch time format, when the environment member last opened the environment.

permissions :: Permissions

The type of environment member permissions associated with this environment member. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

userId :: Text

The user ID in Identity and Access Management (IAM) of the environment member.

userArn :: Text

The Amazon Resource Name (ARN) of the environment member.

environmentId :: Text

The ID of the environment for the environment member.

newEnvironmentMember

Arguments

:: Permissions
$sel:permissions:EnvironmentMember'

-> Text
$sel:userId:EnvironmentMember'

-> Text
$sel:userArn:EnvironmentMember'

-> Text
$sel:environmentId:EnvironmentMember'

-> EnvironmentMember

Create a value of EnvironmentMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastAccess:EnvironmentMember', environmentMember_lastAccess - The time, expressed in epoch time format, when the environment member last opened the environment.

$sel:permissions:EnvironmentMember', environmentMember_permissions - The type of environment member permissions associated with this environment member. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

$sel:userId:EnvironmentMember', environmentMember_userId - The user ID in Identity and Access Management (IAM) of the environment member.

$sel:userArn:EnvironmentMember', environmentMember_userArn - The Amazon Resource Name (ARN) of the environment member.

$sel:environmentId:EnvironmentMember', environmentMember_environmentId - The ID of the environment for the environment member.

environmentMember_lastAccess :: Lens' EnvironmentMember (Maybe UTCTime)

The time, expressed in epoch time format, when the environment member last opened the environment.

environmentMember_permissions :: Lens' EnvironmentMember Permissions

The type of environment member permissions associated with this environment member. Available values include:

owner: Owns the environment.

read-only: Has read-only access to the environment.

read-write: Has read-write access to the environment.

environmentMember_userId :: Lens' EnvironmentMember Text

The user ID in Identity and Access Management (IAM) of the environment member.

environmentMember_userArn :: Lens' EnvironmentMember Text

The Amazon Resource Name (ARN) of the environment member.

environmentMember_environmentId :: Lens' EnvironmentMember Text

The ID of the environment for the environment member.