Network.AWS.CloudControl.ListResourceRequests

Description

Returns existing resource operation requests. This includes requests of all status types. For more information, see Listing active resource operation requests in the Amazon Web Services Cloud Control API User Guide.

Resource operation requests expire after seven days.

Creating a Request

data ListResourceRequests

See: newListResourceRequests smart constructor.

Constructors

ListResourceRequests'

Fields

nextToken :: Maybe Text

If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

resourceRequestStatusFilter :: Maybe ResourceRequestStatusFilter

The filter criteria to apply to the requests returned.

maxResults :: Maybe Natural

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

The default is 20.

newListResourceRequests :: ListResourceRequests

Create a value of ListResourceRequests with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListResourceRequests', listResourceRequests_nextToken - If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:resourceRequestStatusFilter:ListResourceRequests', listResourceRequests_resourceRequestStatusFilter - The filter criteria to apply to the requests returned.

$sel:maxResults:ListResourceRequests', listResourceRequests_maxResults - The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

The default is 20.

Request Lenses

listResourceRequests_nextToken :: Lens' ListResourceRequests (Maybe Text)

If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listResourceRequests_resourceRequestStatusFilter :: Lens' ListResourceRequests (Maybe ResourceRequestStatusFilter)

The filter criteria to apply to the requests returned.

listResourceRequests_maxResults :: Lens' ListResourceRequests (Maybe Natural)

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

The default is 20.

Destructuring the Response

data ListResourceRequestsResponse

See: newListResourceRequestsResponse smart constructor.

Constructors

ListResourceRequestsResponse'

Fields

resourceRequestStatusSummaries :: Maybe [ProgressEvent]

The requests that match the specified filter criteria.

nextToken :: Maybe Text

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListResources again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

httpStatus :: Int

The response's http status code.

newListResourceRequestsResponse

Arguments

:: Int
$sel:httpStatus:ListResourceRequestsResponse'

-> ListResourceRequestsResponse

Create a value of ListResourceRequestsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceRequestStatusSummaries:ListResourceRequestsResponse', listResourceRequestsResponse_resourceRequestStatusSummaries - The requests that match the specified filter criteria.

$sel:nextToken:ListResourceRequests', listResourceRequestsResponse_nextToken - If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListResources again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

$sel:httpStatus:ListResourceRequestsResponse', listResourceRequestsResponse_httpStatus - The response's http status code.

Response Lenses

listResourceRequestsResponse_resourceRequestStatusSummaries :: Lens' ListResourceRequestsResponse (Maybe [ProgressEvent])

The requests that match the specified filter criteria.

listResourceRequestsResponse_nextToken :: Lens' ListResourceRequestsResponse (Maybe Text)

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListResources again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

listResourceRequestsResponse_httpStatus :: Lens' ListResourceRequestsResponse Int

The response's http status code.