Network.AWS.CloudDirectory.AttachTypedLink

Description

Attaches a typed link to a specified source and target object. For more information, see Typed Links.

Creating a Request

data AttachTypedLink

See: newAttachTypedLink smart constructor.

Constructors

AttachTypedLink'

Fields

directoryArn :: Text

The Amazon Resource Name (ARN) of the directory where you want to attach the typed link.

sourceObjectReference :: ObjectReference

Identifies the source object that the typed link will attach to.

targetObjectReference :: ObjectReference

Identifies the target object that the typed link will attach to.

typedLinkFacet :: TypedLinkSchemaAndFacetName

Identifies the typed link facet that is associated with the typed link.

attributes :: [AttributeNameAndValue]

A set of attributes that are associated with the typed link.

newAttachTypedLink

Arguments

:: Text
$sel:directoryArn:AttachTypedLink'

-> ObjectReference
$sel:sourceObjectReference:AttachTypedLink'

-> ObjectReference
$sel:targetObjectReference:AttachTypedLink'

-> TypedLinkSchemaAndFacetName
$sel:typedLinkFacet:AttachTypedLink'

-> AttachTypedLink

Create a value of AttachTypedLink with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryArn:AttachTypedLink', attachTypedLink_directoryArn - The Amazon Resource Name (ARN) of the directory where you want to attach the typed link.

$sel:sourceObjectReference:AttachTypedLink', attachTypedLink_sourceObjectReference - Identifies the source object that the typed link will attach to.

$sel:targetObjectReference:AttachTypedLink', attachTypedLink_targetObjectReference - Identifies the target object that the typed link will attach to.

$sel:typedLinkFacet:AttachTypedLink', attachTypedLink_typedLinkFacet - Identifies the typed link facet that is associated with the typed link.

$sel:attributes:AttachTypedLink', attachTypedLink_attributes - A set of attributes that are associated with the typed link.

Request Lenses

attachTypedLink_directoryArn :: Lens' AttachTypedLink Text

The Amazon Resource Name (ARN) of the directory where you want to attach the typed link.

attachTypedLink_sourceObjectReference :: Lens' AttachTypedLink ObjectReference

Identifies the source object that the typed link will attach to.

attachTypedLink_targetObjectReference :: Lens' AttachTypedLink ObjectReference

Identifies the target object that the typed link will attach to.

attachTypedLink_typedLinkFacet :: Lens' AttachTypedLink TypedLinkSchemaAndFacetName

Identifies the typed link facet that is associated with the typed link.

attachTypedLink_attributes :: Lens' AttachTypedLink [AttributeNameAndValue]

A set of attributes that are associated with the typed link.

Destructuring the Response

data AttachTypedLinkResponse

See: newAttachTypedLinkResponse smart constructor.

Constructors

AttachTypedLinkResponse'

Fields

typedLinkSpecifier :: Maybe TypedLinkSpecifier

Returns a typed link specifier as output.

httpStatus :: Int

The response's http status code.

newAttachTypedLinkResponse

Arguments

:: Int
$sel:httpStatus:AttachTypedLinkResponse'

-> AttachTypedLinkResponse

Create a value of AttachTypedLinkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typedLinkSpecifier:AttachTypedLinkResponse', attachTypedLinkResponse_typedLinkSpecifier - Returns a typed link specifier as output.

$sel:httpStatus:AttachTypedLinkResponse', attachTypedLinkResponse_httpStatus - The response's http status code.

Response Lenses

attachTypedLinkResponse_typedLinkSpecifier :: Lens' AttachTypedLinkResponse (Maybe TypedLinkSpecifier)

Returns a typed link specifier as output.

attachTypedLinkResponse_httpStatus :: Lens' AttachTypedLinkResponse Int

The response's http status code.