Network.AWS.CloudDirectory.CreateDirectory

Description

Creates a Directory by copying the published schema into the directory. A directory cannot be created without a schema.

You can also quickly create a directory using a managed schema, called the QuickStartSchema. For more information, see Managed Schema in the Amazon Cloud Directory Developer Guide.

Creating a Request

data CreateDirectory

See: newCreateDirectory smart constructor.

Constructors

CreateDirectory'

Fields

name :: Text

The name of the Directory. Should be unique per account, per region.

schemaArn :: Text

The Amazon Resource Name (ARN) of the published schema that will be copied into the data Directory. For more information, see arns.

newCreateDirectory

Arguments

:: Text
$sel:name:CreateDirectory'

-> Text
$sel:schemaArn:CreateDirectory'

-> CreateDirectory

Create a value of CreateDirectory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateDirectory', createDirectory_name - The name of the Directory. Should be unique per account, per region.

$sel:schemaArn:CreateDirectory', createDirectory_schemaArn - The Amazon Resource Name (ARN) of the published schema that will be copied into the data Directory. For more information, see arns.

Request Lenses

createDirectory_name :: Lens' CreateDirectory Text

The name of the Directory. Should be unique per account, per region.

createDirectory_schemaArn :: Lens' CreateDirectory Text

The Amazon Resource Name (ARN) of the published schema that will be copied into the data Directory. For more information, see arns.

Destructuring the Response

data CreateDirectoryResponse

See: newCreateDirectoryResponse smart constructor.

Constructors

CreateDirectoryResponse'

Fields

httpStatus :: Int

The response's http status code.

directoryArn :: Text

The ARN that is associated with the Directory. For more information, see arns.

name :: Text

The name of the Directory.

objectIdentifier :: Text

The root object node of the created directory.

appliedSchemaArn :: Text

The ARN of the published schema in the Directory. Once a published schema is copied into the directory, it has its own ARN, which is referred to applied schema ARN. For more information, see arns.

newCreateDirectoryResponse

Arguments

:: Int
$sel:httpStatus:CreateDirectoryResponse'

-> Text
$sel:directoryArn:CreateDirectoryResponse'

-> Text
$sel:name:CreateDirectory'

-> Text
$sel:objectIdentifier:CreateDirectoryResponse'

-> Text
$sel:appliedSchemaArn:CreateDirectoryResponse'

-> CreateDirectoryResponse

Create a value of CreateDirectoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateDirectoryResponse', createDirectoryResponse_httpStatus - The response's http status code.

$sel:directoryArn:CreateDirectoryResponse', createDirectoryResponse_directoryArn - The ARN that is associated with the Directory. For more information, see arns.

$sel:name:CreateDirectory', createDirectoryResponse_name - The name of the Directory.

$sel:objectIdentifier:CreateDirectoryResponse', createDirectoryResponse_objectIdentifier - The root object node of the created directory.

$sel:appliedSchemaArn:CreateDirectoryResponse', createDirectoryResponse_appliedSchemaArn - The ARN of the published schema in the Directory. Once a published schema is copied into the directory, it has its own ARN, which is referred to applied schema ARN. For more information, see arns.

Response Lenses

createDirectoryResponse_httpStatus :: Lens' CreateDirectoryResponse Int

The response's http status code.

createDirectoryResponse_directoryArn :: Lens' CreateDirectoryResponse Text

The ARN that is associated with the Directory. For more information, see arns.

createDirectoryResponse_name :: Lens' CreateDirectoryResponse Text

The name of the Directory.

createDirectoryResponse_objectIdentifier :: Lens' CreateDirectoryResponse Text

The root object node of the created directory.

createDirectoryResponse_appliedSchemaArn :: Lens' CreateDirectoryResponse Text

The ARN of the published schema in the Directory. Once a published schema is copied into the directory, it has its own ARN, which is referred to applied schema ARN. For more information, see arns.