Network.AWS.CloudDirectory.CreateIndex

Description

Creates an index object. See Indexing and search for more information.

Creating a Request

data CreateIndex

See: newCreateIndex smart constructor.

Constructors

CreateIndex'

Fields

linkName :: Maybe Text

The name of the link between the parent object and the index object.

parentReference :: Maybe ObjectReference

A reference to the parent object that contains the index object.

directoryArn :: Text

The ARN of the directory where the index should be created.

orderedIndexedAttributeList :: [AttributeKey]

Specifies the attributes that should be indexed on. Currently only a single attribute is supported.

isUnique :: Bool

Indicates whether the attribute that is being indexed has unique values or not.

newCreateIndex

Arguments

:: Text
$sel:directoryArn:CreateIndex'

-> Bool
$sel:isUnique:CreateIndex'

-> CreateIndex

Create a value of CreateIndex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:linkName:CreateIndex', createIndex_linkName - The name of the link between the parent object and the index object.

$sel:parentReference:CreateIndex', createIndex_parentReference - A reference to the parent object that contains the index object.

$sel:directoryArn:CreateIndex', createIndex_directoryArn - The ARN of the directory where the index should be created.

$sel:orderedIndexedAttributeList:CreateIndex', createIndex_orderedIndexedAttributeList - Specifies the attributes that should be indexed on. Currently only a single attribute is supported.

$sel:isUnique:CreateIndex', createIndex_isUnique - Indicates whether the attribute that is being indexed has unique values or not.

Request Lenses

createIndex_linkName :: Lens' CreateIndex (Maybe Text)

The name of the link between the parent object and the index object.

createIndex_parentReference :: Lens' CreateIndex (Maybe ObjectReference)

A reference to the parent object that contains the index object.

createIndex_directoryArn :: Lens' CreateIndex Text

The ARN of the directory where the index should be created.

createIndex_orderedIndexedAttributeList :: Lens' CreateIndex [AttributeKey]

Specifies the attributes that should be indexed on. Currently only a single attribute is supported.

createIndex_isUnique :: Lens' CreateIndex Bool

Indicates whether the attribute that is being indexed has unique values or not.

Destructuring the Response

data CreateIndexResponse

See: newCreateIndexResponse smart constructor.

Constructors

CreateIndexResponse'

Fields

objectIdentifier :: Maybe Text

The ObjectIdentifier of the index created by this operation.

httpStatus :: Int

The response's http status code.

newCreateIndexResponse

Arguments

:: Int
$sel:httpStatus:CreateIndexResponse'

-> CreateIndexResponse

Create a value of CreateIndexResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:objectIdentifier:CreateIndexResponse', createIndexResponse_objectIdentifier - The ObjectIdentifier of the index created by this operation.

$sel:httpStatus:CreateIndexResponse', createIndexResponse_httpStatus - The response's http status code.

Response Lenses

createIndexResponse_objectIdentifier :: Lens' CreateIndexResponse (Maybe Text)

The ObjectIdentifier of the index created by this operation.

createIndexResponse_httpStatus :: Lens' CreateIndexResponse Int

The response's http status code.