Network.AWS.CloudDirectory.CreateObject

Description

Creates an object in a Directory. Additionally attaches the object to a parent, if a parent reference and LinkName is specified. An object is simply a collection of Facet attributes. You can also use this API call to create a policy object, if the facet from which you create the object is a policy facet.

Creating a Request

data CreateObject

See: newCreateObject smart constructor.

Constructors

CreateObject'

Fields

linkName :: Maybe Text

The name of link that is used to attach this object to a parent.

parentReference :: Maybe ObjectReference

If specified, the parent reference to which this object will be attached.

objectAttributeList :: Maybe [AttributeKeyAndValue]

The attribute map whose attribute ARN contains the key and attribute value as the map value.

directoryArn :: Text

The Amazon Resource Name (ARN) that is associated with the Directory in which the object will be created. For more information, see arns.

schemaFacets :: [SchemaFacet]

A list of schema facets to be associated with the object. Do not provide minor version components. See SchemaFacet for details.

newCreateObject

Arguments

:: Text
$sel:directoryArn:CreateObject'

-> CreateObject

Create a value of CreateObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:linkName:CreateObject', createObject_linkName - The name of link that is used to attach this object to a parent.

$sel:parentReference:CreateObject', createObject_parentReference - If specified, the parent reference to which this object will be attached.

$sel:objectAttributeList:CreateObject', createObject_objectAttributeList - The attribute map whose attribute ARN contains the key and attribute value as the map value.

$sel:directoryArn:CreateObject', createObject_directoryArn - The Amazon Resource Name (ARN) that is associated with the Directory in which the object will be created. For more information, see arns.

$sel:schemaFacets:CreateObject', createObject_schemaFacets - A list of schema facets to be associated with the object. Do not provide minor version components. See SchemaFacet for details.

Request Lenses

createObject_linkName :: Lens' CreateObject (Maybe Text)

The name of link that is used to attach this object to a parent.

createObject_parentReference :: Lens' CreateObject (Maybe ObjectReference)

If specified, the parent reference to which this object will be attached.

createObject_objectAttributeList :: Lens' CreateObject (Maybe [AttributeKeyAndValue])

The attribute map whose attribute ARN contains the key and attribute value as the map value.

createObject_directoryArn :: Lens' CreateObject Text

The Amazon Resource Name (ARN) that is associated with the Directory in which the object will be created. For more information, see arns.

createObject_schemaFacets :: Lens' CreateObject [SchemaFacet]

A list of schema facets to be associated with the object. Do not provide minor version components. See SchemaFacet for details.

Destructuring the Response

data CreateObjectResponse

See: newCreateObjectResponse smart constructor.

Constructors

CreateObjectResponse'

Fields

objectIdentifier :: Maybe Text

The identifier that is associated with the object.

httpStatus :: Int

The response's http status code.

newCreateObjectResponse

Arguments

:: Int
$sel:httpStatus:CreateObjectResponse'

-> CreateObjectResponse

Create a value of CreateObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:objectIdentifier:CreateObjectResponse', createObjectResponse_objectIdentifier - The identifier that is associated with the object.

$sel:httpStatus:CreateObjectResponse', createObjectResponse_httpStatus - The response's http status code.

Response Lenses

createObjectResponse_objectIdentifier :: Lens' CreateObjectResponse (Maybe Text)

The identifier that is associated with the object.

createObjectResponse_httpStatus :: Lens' CreateObjectResponse Int

The response's http status code.