Network.AWS.CloudDirectory.CreateTypedLinkFacet

Description

Creates a TypedLinkFacet. For more information, see Typed Links.

Creating a Request

data CreateTypedLinkFacet

See: newCreateTypedLinkFacet smart constructor.

Constructors

CreateTypedLinkFacet'

Fields

schemaArn :: Text

The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns.

facet :: TypedLinkFacet

Facet structure that is associated with the typed link facet.

newCreateTypedLinkFacet

Arguments

:: Text
$sel:schemaArn:CreateTypedLinkFacet'

-> TypedLinkFacet
$sel:facet:CreateTypedLinkFacet'

-> CreateTypedLinkFacet

Create a value of CreateTypedLinkFacet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaArn:CreateTypedLinkFacet', createTypedLinkFacet_schemaArn - The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns.

$sel:facet:CreateTypedLinkFacet', createTypedLinkFacet_facet - Facet structure that is associated with the typed link facet.

Request Lenses

createTypedLinkFacet_schemaArn :: Lens' CreateTypedLinkFacet Text

The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns.

createTypedLinkFacet_facet :: Lens' CreateTypedLinkFacet TypedLinkFacet

Facet structure that is associated with the typed link facet.

Destructuring the Response

data CreateTypedLinkFacetResponse

See: newCreateTypedLinkFacetResponse smart constructor.

Constructors

CreateTypedLinkFacetResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateTypedLinkFacetResponse

Arguments

:: Int
$sel:httpStatus:CreateTypedLinkFacetResponse'

-> CreateTypedLinkFacetResponse

Create a value of CreateTypedLinkFacetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateTypedLinkFacetResponse', createTypedLinkFacetResponse_httpStatus - The response's http status code.

Response Lenses

createTypedLinkFacetResponse_httpStatus :: Lens' CreateTypedLinkFacetResponse Int

The response's http status code.