Network.AWS.CloudDirectory.GetSchemaAsJson

Description

Retrieves a JSON representation of the schema. See JSON Schema Format for more information.

Creating a Request

data GetSchemaAsJson

See: newGetSchemaAsJson smart constructor.

Constructors

GetSchemaAsJson'

Fields

schemaArn :: Text

The ARN of the schema to retrieve.

newGetSchemaAsJson

Arguments

:: Text
$sel:schemaArn:GetSchemaAsJson'

-> GetSchemaAsJson

Create a value of GetSchemaAsJson with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaArn:GetSchemaAsJson', getSchemaAsJson_schemaArn - The ARN of the schema to retrieve.

Request Lenses

getSchemaAsJson_schemaArn :: Lens' GetSchemaAsJson Text

The ARN of the schema to retrieve.

Destructuring the Response

data GetSchemaAsJsonResponse

See: newGetSchemaAsJsonResponse smart constructor.

Constructors

GetSchemaAsJsonResponse'

Fields

name :: Maybe Text

The name of the retrieved schema.

document :: Maybe Text

The JSON representation of the schema document.

httpStatus :: Int

The response's http status code.

newGetSchemaAsJsonResponse

Arguments

:: Int
$sel:httpStatus:GetSchemaAsJsonResponse'

-> GetSchemaAsJsonResponse

Create a value of GetSchemaAsJsonResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetSchemaAsJsonResponse', getSchemaAsJsonResponse_name - The name of the retrieved schema.

$sel:document:GetSchemaAsJsonResponse', getSchemaAsJsonResponse_document - The JSON representation of the schema document.

$sel:httpStatus:GetSchemaAsJsonResponse', getSchemaAsJsonResponse_httpStatus - The response's http status code.

Response Lenses

getSchemaAsJsonResponse_name :: Lens' GetSchemaAsJsonResponse (Maybe Text)

The name of the retrieved schema.

getSchemaAsJsonResponse_document :: Lens' GetSchemaAsJsonResponse (Maybe Text)

The JSON representation of the schema document.

getSchemaAsJsonResponse_httpStatus :: Lens' GetSchemaAsJsonResponse Int

The response's http status code.