Network.AWS.CloudDirectory.ListFacetNames

Description

Retrieves the names of facets that exist in a schema.

This operation returns paginated results.

Creating a Request

data ListFacetNames

See: newListFacetNames smart constructor.

Constructors

ListFacetNames'

Fields

nextToken :: Maybe Text

The pagination token.

maxResults :: Maybe Natural

The maximum number of results to retrieve.

schemaArn :: Text

The Amazon Resource Name (ARN) to retrieve facet names from.

newListFacetNames

Arguments

:: Text
$sel:schemaArn:ListFacetNames'

-> ListFacetNames

Create a value of ListFacetNames with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFacetNames', listFacetNames_nextToken - The pagination token.

$sel:maxResults:ListFacetNames', listFacetNames_maxResults - The maximum number of results to retrieve.

$sel:schemaArn:ListFacetNames', listFacetNames_schemaArn - The Amazon Resource Name (ARN) to retrieve facet names from.

Request Lenses

listFacetNames_nextToken :: Lens' ListFacetNames (Maybe Text)

The pagination token.

listFacetNames_maxResults :: Lens' ListFacetNames (Maybe Natural)

The maximum number of results to retrieve.

listFacetNames_schemaArn :: Lens' ListFacetNames Text

The Amazon Resource Name (ARN) to retrieve facet names from.

Destructuring the Response

data ListFacetNamesResponse

See: newListFacetNamesResponse smart constructor.

Constructors

ListFacetNamesResponse'

Fields

nextToken :: Maybe Text

The pagination token.

facetNames :: Maybe [Text]

The names of facets that exist within the schema.

httpStatus :: Int

The response's http status code.

newListFacetNamesResponse

Arguments

:: Int
$sel:httpStatus:ListFacetNamesResponse'

-> ListFacetNamesResponse

Create a value of ListFacetNamesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFacetNames', listFacetNamesResponse_nextToken - The pagination token.

$sel:facetNames:ListFacetNamesResponse', listFacetNamesResponse_facetNames - The names of facets that exist within the schema.

$sel:httpStatus:ListFacetNamesResponse', listFacetNamesResponse_httpStatus - The response's http status code.

Response Lenses

listFacetNamesResponse_nextToken :: Lens' ListFacetNamesResponse (Maybe Text)

The pagination token.

listFacetNamesResponse_facetNames :: Lens' ListFacetNamesResponse (Maybe [Text])

The names of facets that exist within the schema.

listFacetNamesResponse_httpStatus :: Lens' ListFacetNamesResponse Int

The response's http status code.