Network.AWS.CloudDirectory.ListIncomingTypedLinks

Description

Returns a paginated list of all the incoming TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

This operation returns paginated results.

Creating a Request

data ListIncomingTypedLinks

See: newListIncomingTypedLinks smart constructor.

Constructors

ListIncomingTypedLinks'

Fields

nextToken :: Maybe Text

The pagination token.

filterTypedLink :: Maybe TypedLinkSchemaAndFacetName

Filters are interpreted in the order of the attributes on the typed link facet, not the order in which they are supplied to any API calls.

maxResults :: Maybe Natural

The maximum number of results to retrieve.

consistencyLevel :: Maybe ConsistencyLevel

The consistency level to execute the request at.

filterAttributeRanges :: Maybe [TypedLinkAttributeRange]

Provides range filters for multiple attributes. When providing ranges to typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range.

directoryArn :: Text

The Amazon Resource Name (ARN) of the directory where you want to list the typed links.

objectReference :: ObjectReference

Reference that identifies the object whose attributes will be listed.

newListIncomingTypedLinks

Arguments

:: Text
$sel:directoryArn:ListIncomingTypedLinks'

-> ObjectReference
$sel:objectReference:ListIncomingTypedLinks'

-> ListIncomingTypedLinks

Create a value of ListIncomingTypedLinks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListIncomingTypedLinks', listIncomingTypedLinks_nextToken - The pagination token.

$sel:filterTypedLink:ListIncomingTypedLinks', listIncomingTypedLinks_filterTypedLink - Filters are interpreted in the order of the attributes on the typed link facet, not the order in which they are supplied to any API calls.

$sel:maxResults:ListIncomingTypedLinks', listIncomingTypedLinks_maxResults - The maximum number of results to retrieve.

$sel:consistencyLevel:ListIncomingTypedLinks', listIncomingTypedLinks_consistencyLevel - The consistency level to execute the request at.

$sel:filterAttributeRanges:ListIncomingTypedLinks', listIncomingTypedLinks_filterAttributeRanges - Provides range filters for multiple attributes. When providing ranges to typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range.

$sel:directoryArn:ListIncomingTypedLinks', listIncomingTypedLinks_directoryArn - The Amazon Resource Name (ARN) of the directory where you want to list the typed links.

$sel:objectReference:ListIncomingTypedLinks', listIncomingTypedLinks_objectReference - Reference that identifies the object whose attributes will be listed.

Request Lenses

listIncomingTypedLinks_nextToken :: Lens' ListIncomingTypedLinks (Maybe Text)

The pagination token.

listIncomingTypedLinks_filterTypedLink :: Lens' ListIncomingTypedLinks (Maybe TypedLinkSchemaAndFacetName)

Filters are interpreted in the order of the attributes on the typed link facet, not the order in which they are supplied to any API calls.

listIncomingTypedLinks_maxResults :: Lens' ListIncomingTypedLinks (Maybe Natural)

The maximum number of results to retrieve.

listIncomingTypedLinks_consistencyLevel :: Lens' ListIncomingTypedLinks (Maybe ConsistencyLevel)

The consistency level to execute the request at.

listIncomingTypedLinks_filterAttributeRanges :: Lens' ListIncomingTypedLinks (Maybe [TypedLinkAttributeRange])

Provides range filters for multiple attributes. When providing ranges to typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range.

listIncomingTypedLinks_directoryArn :: Lens' ListIncomingTypedLinks Text

The Amazon Resource Name (ARN) of the directory where you want to list the typed links.

listIncomingTypedLinks_objectReference :: Lens' ListIncomingTypedLinks ObjectReference

Reference that identifies the object whose attributes will be listed.

Destructuring the Response

data ListIncomingTypedLinksResponse

See: newListIncomingTypedLinksResponse smart constructor.

Constructors

ListIncomingTypedLinksResponse'

Fields

linkSpecifiers :: Maybe [TypedLinkSpecifier]

Returns one or more typed link specifiers as output.

nextToken :: Maybe Text

The pagination token.

httpStatus :: Int

The response's http status code.

newListIncomingTypedLinksResponse

Arguments

:: Int
$sel:httpStatus:ListIncomingTypedLinksResponse'

-> ListIncomingTypedLinksResponse

Create a value of ListIncomingTypedLinksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:linkSpecifiers:ListIncomingTypedLinksResponse', listIncomingTypedLinksResponse_linkSpecifiers - Returns one or more typed link specifiers as output.

$sel:nextToken:ListIncomingTypedLinks', listIncomingTypedLinksResponse_nextToken - The pagination token.

$sel:httpStatus:ListIncomingTypedLinksResponse', listIncomingTypedLinksResponse_httpStatus - The response's http status code.

Response Lenses

listIncomingTypedLinksResponse_linkSpecifiers :: Lens' ListIncomingTypedLinksResponse (Maybe [TypedLinkSpecifier])

Returns one or more typed link specifiers as output.

listIncomingTypedLinksResponse_nextToken :: Lens' ListIncomingTypedLinksResponse (Maybe Text)

The pagination token.

listIncomingTypedLinksResponse_httpStatus :: Lens' ListIncomingTypedLinksResponse Int

The response's http status code.