Network.AWS.CloudDirectory.ListObjectAttributes

Description

Lists all attributes that are associated with an object.

This operation returns paginated results.

Creating a Request

data ListObjectAttributes

See: newListObjectAttributes smart constructor.

Constructors

ListObjectAttributes'

Fields

facetFilter :: Maybe SchemaFacet

Used to filter the list of object attributes that are associated with a certain facet.

consistencyLevel :: Maybe ConsistencyLevel

Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

nextToken :: Maybe Text

The pagination token.

maxResults :: Maybe Natural

The maximum number of items to be retrieved in a single call. This is an approximate number.

directoryArn :: Text

The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns.

objectReference :: ObjectReference

The reference that identifies the object whose attributes will be listed.

newListObjectAttributes

Arguments

:: Text
$sel:directoryArn:ListObjectAttributes'

-> ObjectReference
$sel:objectReference:ListObjectAttributes'

-> ListObjectAttributes

Create a value of ListObjectAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:facetFilter:ListObjectAttributes', listObjectAttributes_facetFilter - Used to filter the list of object attributes that are associated with a certain facet.

$sel:consistencyLevel:ListObjectAttributes', listObjectAttributes_consistencyLevel - Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

$sel:nextToken:ListObjectAttributes', listObjectAttributes_nextToken - The pagination token.

$sel:maxResults:ListObjectAttributes', listObjectAttributes_maxResults - The maximum number of items to be retrieved in a single call. This is an approximate number.

$sel:directoryArn:ListObjectAttributes', listObjectAttributes_directoryArn - The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns.

$sel:objectReference:ListObjectAttributes', listObjectAttributes_objectReference - The reference that identifies the object whose attributes will be listed.

Request Lenses

listObjectAttributes_facetFilter :: Lens' ListObjectAttributes (Maybe SchemaFacet)

Used to filter the list of object attributes that are associated with a certain facet.

listObjectAttributes_consistencyLevel :: Lens' ListObjectAttributes (Maybe ConsistencyLevel)

Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

listObjectAttributes_nextToken :: Lens' ListObjectAttributes (Maybe Text)

The pagination token.

listObjectAttributes_maxResults :: Lens' ListObjectAttributes (Maybe Natural)

The maximum number of items to be retrieved in a single call. This is an approximate number.

listObjectAttributes_directoryArn :: Lens' ListObjectAttributes Text

The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns.

listObjectAttributes_objectReference :: Lens' ListObjectAttributes ObjectReference

The reference that identifies the object whose attributes will be listed.

Destructuring the Response

data ListObjectAttributesResponse

See: newListObjectAttributesResponse smart constructor.

Constructors

ListObjectAttributesResponse'

Fields

nextToken :: Maybe Text

The pagination token.

attributes :: Maybe [AttributeKeyAndValue]

Attributes map that is associated with the object. AttributeArn is the key, and attribute value is the value.

httpStatus :: Int

The response's http status code.

newListObjectAttributesResponse

Arguments

:: Int
$sel:httpStatus:ListObjectAttributesResponse'

-> ListObjectAttributesResponse

Create a value of ListObjectAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListObjectAttributes', listObjectAttributesResponse_nextToken - The pagination token.

$sel:attributes:ListObjectAttributesResponse', listObjectAttributesResponse_attributes - Attributes map that is associated with the object. AttributeArn is the key, and attribute value is the value.

$sel:httpStatus:ListObjectAttributesResponse', listObjectAttributesResponse_httpStatus - The response's http status code.

Response Lenses

listObjectAttributesResponse_nextToken :: Lens' ListObjectAttributesResponse (Maybe Text)

The pagination token.

listObjectAttributesResponse_attributes :: Lens' ListObjectAttributesResponse (Maybe [AttributeKeyAndValue])

Attributes map that is associated with the object. AttributeArn is the key, and attribute value is the value.

listObjectAttributesResponse_httpStatus :: Lens' ListObjectAttributesResponse Int

The response's http status code.