Network.AWS.CloudDirectory.ListObjectParents

Description

Lists parent objects that are associated with a given object in pagination fashion.

Creating a Request

data ListObjectParents

See: newListObjectParents smart constructor.

Constructors

ListObjectParents'

Fields

nextToken :: Maybe Text

The pagination token.

maxResults :: Maybe Natural

The maximum number of items to be retrieved in a single call. This is an approximate number.

includeAllLinksToEachParent :: Maybe Bool

When set to True, returns all ListObjectParentsResponse$ParentLinks. There could be multiple links between a parent-child pair.

consistencyLevel :: Maybe ConsistencyLevel

Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

directoryArn :: Text

The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns.

objectReference :: ObjectReference

The reference that identifies the object for which parent objects are being listed.

newListObjectParents

Arguments

:: Text
$sel:directoryArn:ListObjectParents'

-> ObjectReference
$sel:objectReference:ListObjectParents'

-> ListObjectParents

Create a value of ListObjectParents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListObjectParents', listObjectParents_nextToken - The pagination token.

$sel:maxResults:ListObjectParents', listObjectParents_maxResults - The maximum number of items to be retrieved in a single call. This is an approximate number.

$sel:includeAllLinksToEachParent:ListObjectParents', listObjectParents_includeAllLinksToEachParent - When set to True, returns all ListObjectParentsResponse$ParentLinks. There could be multiple links between a parent-child pair.

$sel:consistencyLevel:ListObjectParents', listObjectParents_consistencyLevel - Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

$sel:directoryArn:ListObjectParents', listObjectParents_directoryArn - The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns.

$sel:objectReference:ListObjectParents', listObjectParents_objectReference - The reference that identifies the object for which parent objects are being listed.

Request Lenses

listObjectParents_nextToken :: Lens' ListObjectParents (Maybe Text)

The pagination token.

listObjectParents_maxResults :: Lens' ListObjectParents (Maybe Natural)

The maximum number of items to be retrieved in a single call. This is an approximate number.

listObjectParents_includeAllLinksToEachParent :: Lens' ListObjectParents (Maybe Bool)

When set to True, returns all ListObjectParentsResponse$ParentLinks. There could be multiple links between a parent-child pair.

listObjectParents_consistencyLevel :: Lens' ListObjectParents (Maybe ConsistencyLevel)

Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

listObjectParents_directoryArn :: Lens' ListObjectParents Text

The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns.

listObjectParents_objectReference :: Lens' ListObjectParents ObjectReference

The reference that identifies the object for which parent objects are being listed.

Destructuring the Response

data ListObjectParentsResponse

See: newListObjectParentsResponse smart constructor.

Constructors

ListObjectParentsResponse'

Fields

parentLinks :: Maybe [ObjectIdentifierAndLinkNameTuple]

Returns a list of parent reference and LinkName Tuples.

parents :: Maybe (HashMap Text Text)

The parent structure, which is a map with key as the ObjectIdentifier and LinkName as the value.

nextToken :: Maybe Text

The pagination token.

httpStatus :: Int

The response's http status code.

newListObjectParentsResponse

Arguments

:: Int
$sel:httpStatus:ListObjectParentsResponse'

-> ListObjectParentsResponse

Create a value of ListObjectParentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentLinks:ListObjectParentsResponse', listObjectParentsResponse_parentLinks - Returns a list of parent reference and LinkName Tuples.

$sel:parents:ListObjectParentsResponse', listObjectParentsResponse_parents - The parent structure, which is a map with key as the ObjectIdentifier and LinkName as the value.

$sel:nextToken:ListObjectParents', listObjectParentsResponse_nextToken - The pagination token.

$sel:httpStatus:ListObjectParentsResponse', listObjectParentsResponse_httpStatus - The response's http status code.

Response Lenses

listObjectParentsResponse_parentLinks :: Lens' ListObjectParentsResponse (Maybe [ObjectIdentifierAndLinkNameTuple])

Returns a list of parent reference and LinkName Tuples.

listObjectParentsResponse_parents :: Lens' ListObjectParentsResponse (Maybe (HashMap Text Text))

The parent structure, which is a map with key as the ObjectIdentifier and LinkName as the value.

listObjectParentsResponse_nextToken :: Lens' ListObjectParentsResponse (Maybe Text)

The pagination token.

listObjectParentsResponse_httpStatus :: Lens' ListObjectParentsResponse Int

The response's http status code.