Network.AWS.CloudDirectory.ListObjectPolicies

Description

Returns policies attached to an object in pagination fashion.

This operation returns paginated results.

Creating a Request

data ListObjectPolicies

See: newListObjectPolicies smart constructor.

Constructors

ListObjectPolicies'

Fields

consistencyLevel :: Maybe ConsistencyLevel

Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

nextToken :: Maybe Text

The pagination token.

maxResults :: Maybe Natural

The maximum number of items to be retrieved in a single call. This is an approximate number.

directoryArn :: Text

The Amazon Resource Name (ARN) that is associated with the Directory where objects reside. For more information, see arns.

objectReference :: ObjectReference

Reference that identifies the object for which policies will be listed.

newListObjectPolicies

Arguments

:: Text
$sel:directoryArn:ListObjectPolicies'

-> ObjectReference
$sel:objectReference:ListObjectPolicies'

-> ListObjectPolicies

Create a value of ListObjectPolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:consistencyLevel:ListObjectPolicies', listObjectPolicies_consistencyLevel - Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

$sel:nextToken:ListObjectPolicies', listObjectPolicies_nextToken - The pagination token.

$sel:maxResults:ListObjectPolicies', listObjectPolicies_maxResults - The maximum number of items to be retrieved in a single call. This is an approximate number.

$sel:directoryArn:ListObjectPolicies', listObjectPolicies_directoryArn - The Amazon Resource Name (ARN) that is associated with the Directory where objects reside. For more information, see arns.

$sel:objectReference:ListObjectPolicies', listObjectPolicies_objectReference - Reference that identifies the object for which policies will be listed.

Request Lenses

listObjectPolicies_consistencyLevel :: Lens' ListObjectPolicies (Maybe ConsistencyLevel)

Represents the manner and timing in which the successful write or update of an object is reflected in a subsequent read operation of that same object.

listObjectPolicies_nextToken :: Lens' ListObjectPolicies (Maybe Text)

The pagination token.

listObjectPolicies_maxResults :: Lens' ListObjectPolicies (Maybe Natural)

The maximum number of items to be retrieved in a single call. This is an approximate number.

listObjectPolicies_directoryArn :: Lens' ListObjectPolicies Text

The Amazon Resource Name (ARN) that is associated with the Directory where objects reside. For more information, see arns.

listObjectPolicies_objectReference :: Lens' ListObjectPolicies ObjectReference

Reference that identifies the object for which policies will be listed.

Destructuring the Response

data ListObjectPoliciesResponse

See: newListObjectPoliciesResponse smart constructor.

Constructors

ListObjectPoliciesResponse'

Fields

nextToken :: Maybe Text

The pagination token.

attachedPolicyIds :: Maybe [Text]

A list of policy ObjectIdentifiers, that are attached to the object.

httpStatus :: Int

The response's http status code.

newListObjectPoliciesResponse

Arguments

:: Int
$sel:httpStatus:ListObjectPoliciesResponse'

-> ListObjectPoliciesResponse

Create a value of ListObjectPoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListObjectPolicies', listObjectPoliciesResponse_nextToken - The pagination token.

$sel:attachedPolicyIds:ListObjectPoliciesResponse', listObjectPoliciesResponse_attachedPolicyIds - A list of policy ObjectIdentifiers, that are attached to the object.

$sel:httpStatus:ListObjectPoliciesResponse', listObjectPoliciesResponse_httpStatus - The response's http status code.

Response Lenses

listObjectPoliciesResponse_nextToken :: Lens' ListObjectPoliciesResponse (Maybe Text)

The pagination token.

listObjectPoliciesResponse_attachedPolicyIds :: Lens' ListObjectPoliciesResponse (Maybe [Text])

A list of policy ObjectIdentifiers, that are attached to the object.

listObjectPoliciesResponse_httpStatus :: Lens' ListObjectPoliciesResponse Int

The response's http status code.