Network.AWS.CloudDirectory.ListOutgoingTypedLinks

Description

Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

This operation returns paginated results.

Creating a Request

data ListOutgoingTypedLinks

See: newListOutgoingTypedLinks smart constructor.

Constructors

ListOutgoingTypedLinks'

Fields

nextToken :: Maybe Text

The pagination token.

filterTypedLink :: Maybe TypedLinkSchemaAndFacetName

Filters are interpreted in the order of the attributes defined on the typed link facet, not the order they are supplied to any API calls.

maxResults :: Maybe Natural

The maximum number of results to retrieve.

consistencyLevel :: Maybe ConsistencyLevel

The consistency level to execute the request at.

filterAttributeRanges :: Maybe [TypedLinkAttributeRange]

Provides range filters for multiple attributes. When providing ranges to typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range.

directoryArn :: Text

The Amazon Resource Name (ARN) of the directory where you want to list the typed links.

objectReference :: ObjectReference

A reference that identifies the object whose attributes will be listed.

newListOutgoingTypedLinks

Arguments

:: Text
$sel:directoryArn:ListOutgoingTypedLinks'

-> ObjectReference
$sel:objectReference:ListOutgoingTypedLinks'

-> ListOutgoingTypedLinks

Create a value of ListOutgoingTypedLinks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListOutgoingTypedLinks', listOutgoingTypedLinks_nextToken - The pagination token.

$sel:filterTypedLink:ListOutgoingTypedLinks', listOutgoingTypedLinks_filterTypedLink - Filters are interpreted in the order of the attributes defined on the typed link facet, not the order they are supplied to any API calls.

$sel:maxResults:ListOutgoingTypedLinks', listOutgoingTypedLinks_maxResults - The maximum number of results to retrieve.

$sel:consistencyLevel:ListOutgoingTypedLinks', listOutgoingTypedLinks_consistencyLevel - The consistency level to execute the request at.

$sel:filterAttributeRanges:ListOutgoingTypedLinks', listOutgoingTypedLinks_filterAttributeRanges - Provides range filters for multiple attributes. When providing ranges to typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range.

$sel:directoryArn:ListOutgoingTypedLinks', listOutgoingTypedLinks_directoryArn - The Amazon Resource Name (ARN) of the directory where you want to list the typed links.

$sel:objectReference:ListOutgoingTypedLinks', listOutgoingTypedLinks_objectReference - A reference that identifies the object whose attributes will be listed.

Request Lenses

listOutgoingTypedLinks_nextToken :: Lens' ListOutgoingTypedLinks (Maybe Text)

The pagination token.

listOutgoingTypedLinks_filterTypedLink :: Lens' ListOutgoingTypedLinks (Maybe TypedLinkSchemaAndFacetName)

Filters are interpreted in the order of the attributes defined on the typed link facet, not the order they are supplied to any API calls.

listOutgoingTypedLinks_maxResults :: Lens' ListOutgoingTypedLinks (Maybe Natural)

The maximum number of results to retrieve.

listOutgoingTypedLinks_consistencyLevel :: Lens' ListOutgoingTypedLinks (Maybe ConsistencyLevel)

The consistency level to execute the request at.

listOutgoingTypedLinks_filterAttributeRanges :: Lens' ListOutgoingTypedLinks (Maybe [TypedLinkAttributeRange])

Provides range filters for multiple attributes. When providing ranges to typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range.

listOutgoingTypedLinks_directoryArn :: Lens' ListOutgoingTypedLinks Text

The Amazon Resource Name (ARN) of the directory where you want to list the typed links.

listOutgoingTypedLinks_objectReference :: Lens' ListOutgoingTypedLinks ObjectReference

A reference that identifies the object whose attributes will be listed.

Destructuring the Response

data ListOutgoingTypedLinksResponse

See: newListOutgoingTypedLinksResponse smart constructor.

Constructors

ListOutgoingTypedLinksResponse'

Fields

nextToken :: Maybe Text

The pagination token.

typedLinkSpecifiers :: Maybe [TypedLinkSpecifier]

Returns a typed link specifier as output.

httpStatus :: Int

The response's http status code.

newListOutgoingTypedLinksResponse

Arguments

:: Int
$sel:httpStatus:ListOutgoingTypedLinksResponse'

-> ListOutgoingTypedLinksResponse

Create a value of ListOutgoingTypedLinksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListOutgoingTypedLinks', listOutgoingTypedLinksResponse_nextToken - The pagination token.

$sel:typedLinkSpecifiers:ListOutgoingTypedLinksResponse', listOutgoingTypedLinksResponse_typedLinkSpecifiers - Returns a typed link specifier as output.

$sel:httpStatus:ListOutgoingTypedLinksResponse', listOutgoingTypedLinksResponse_httpStatus - The response's http status code.

Response Lenses

listOutgoingTypedLinksResponse_nextToken :: Lens' ListOutgoingTypedLinksResponse (Maybe Text)

The pagination token.

listOutgoingTypedLinksResponse_typedLinkSpecifiers :: Lens' ListOutgoingTypedLinksResponse (Maybe [TypedLinkSpecifier])

Returns a typed link specifier as output.

listOutgoingTypedLinksResponse_httpStatus :: Lens' ListOutgoingTypedLinksResponse Int

The response's http status code.