Network.AWS.CloudDirectory.PublishSchema

Description

Publishes a development schema with a major version and a recommended minor version.

Creating a Request

data PublishSchema

See: newPublishSchema smart constructor.

Constructors

PublishSchema'

Fields

name :: Maybe Text

The new name under which the schema will be published. If this is not provided, the development schema is considered.

minorVersion :: Maybe Text

The minor version under which the schema will be published. This parameter is recommended. Schemas have both a major and minor version associated with them.

developmentSchemaArn :: Text

The Amazon Resource Name (ARN) that is associated with the development schema. For more information, see arns.

version :: Text

The major version under which the schema will be published. Schemas have both a major and minor version associated with them.

newPublishSchema

Arguments

:: Text
$sel:developmentSchemaArn:PublishSchema'

-> Text
$sel:version:PublishSchema'

-> PublishSchema

Create a value of PublishSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:PublishSchema', publishSchema_name - The new name under which the schema will be published. If this is not provided, the development schema is considered.

$sel:minorVersion:PublishSchema', publishSchema_minorVersion - The minor version under which the schema will be published. This parameter is recommended. Schemas have both a major and minor version associated with them.

$sel:developmentSchemaArn:PublishSchema', publishSchema_developmentSchemaArn - The Amazon Resource Name (ARN) that is associated with the development schema. For more information, see arns.

$sel:version:PublishSchema', publishSchema_version - The major version under which the schema will be published. Schemas have both a major and minor version associated with them.

Request Lenses

publishSchema_name :: Lens' PublishSchema (Maybe Text)

The new name under which the schema will be published. If this is not provided, the development schema is considered.

publishSchema_minorVersion :: Lens' PublishSchema (Maybe Text)

The minor version under which the schema will be published. This parameter is recommended. Schemas have both a major and minor version associated with them.

publishSchema_developmentSchemaArn :: Lens' PublishSchema Text

The Amazon Resource Name (ARN) that is associated with the development schema. For more information, see arns.

publishSchema_version :: Lens' PublishSchema Text

The major version under which the schema will be published. Schemas have both a major and minor version associated with them.

Destructuring the Response

data PublishSchemaResponse

See: newPublishSchemaResponse smart constructor.

Constructors

PublishSchemaResponse'

Fields

publishedSchemaArn :: Maybe Text

The ARN that is associated with the published schema. For more information, see arns.

httpStatus :: Int

The response's http status code.

newPublishSchemaResponse

Arguments

:: Int
$sel:httpStatus:PublishSchemaResponse'

-> PublishSchemaResponse

Create a value of PublishSchemaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:publishedSchemaArn:PublishSchemaResponse', publishSchemaResponse_publishedSchemaArn - The ARN that is associated with the published schema. For more information, see arns.

$sel:httpStatus:PublishSchemaResponse', publishSchemaResponse_httpStatus - The response's http status code.

Response Lenses

publishSchemaResponse_publishedSchemaArn :: Lens' PublishSchemaResponse (Maybe Text)

The ARN that is associated with the published schema. For more information, see arns.

publishSchemaResponse_httpStatus :: Lens' PublishSchemaResponse Int

The response's http status code.