Network.AWS.CloudDirectory.Types.BatchAttachObject

Description

Documentation

data BatchAttachObject

Represents the output of an AttachObject operation.

See: newBatchAttachObject smart constructor.

Constructors

BatchAttachObject'

Fields

parentReference :: ObjectReference

The parent object reference.

childReference :: ObjectReference

The child object reference that is to be attached to the object.

linkName :: Text

The name of the link.

newBatchAttachObject

Arguments

:: ObjectReference
$sel:parentReference:BatchAttachObject'

-> ObjectReference
$sel:childReference:BatchAttachObject'

-> Text
$sel:linkName:BatchAttachObject'

-> BatchAttachObject

Create a value of BatchAttachObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentReference:BatchAttachObject', batchAttachObject_parentReference - The parent object reference.

$sel:childReference:BatchAttachObject', batchAttachObject_childReference - The child object reference that is to be attached to the object.

$sel:linkName:BatchAttachObject', batchAttachObject_linkName - The name of the link.

batchAttachObject_parentReference :: Lens' BatchAttachObject ObjectReference

The parent object reference.

batchAttachObject_childReference :: Lens' BatchAttachObject ObjectReference

The child object reference that is to be attached to the object.

batchAttachObject_linkName :: Lens' BatchAttachObject Text

The name of the link.