Network.AWS.CloudDirectory.Types.BatchCreateIndex

Description

Documentation

data BatchCreateIndex

Creates an index object inside of a BatchRead operation. For more information, see CreateIndex and BatchReadRequest$Operations.

See: newBatchCreateIndex smart constructor.

Constructors

BatchCreateIndex'

Fields

parentReference :: Maybe ObjectReference

A reference to the parent object that contains the index object.

linkName :: Maybe Text

The name of the link between the parent object and the index object.

batchReferenceName :: Maybe Text

The batch reference name. See Transaction Support for more information.

orderedIndexedAttributeList :: [AttributeKey]

Specifies the attributes that should be indexed on. Currently only a single attribute is supported.

isUnique :: Bool

Indicates whether the attribute that is being indexed has unique values or not.

newBatchCreateIndex

Arguments

:: Bool
$sel:isUnique:BatchCreateIndex'

-> BatchCreateIndex

Create a value of BatchCreateIndex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentReference:BatchCreateIndex', batchCreateIndex_parentReference - A reference to the parent object that contains the index object.

$sel:linkName:BatchCreateIndex', batchCreateIndex_linkName - The name of the link between the parent object and the index object.

$sel:batchReferenceName:BatchCreateIndex', batchCreateIndex_batchReferenceName - The batch reference name. See Transaction Support for more information.

$sel:orderedIndexedAttributeList:BatchCreateIndex', batchCreateIndex_orderedIndexedAttributeList - Specifies the attributes that should be indexed on. Currently only a single attribute is supported.

$sel:isUnique:BatchCreateIndex', batchCreateIndex_isUnique - Indicates whether the attribute that is being indexed has unique values or not.

batchCreateIndex_parentReference :: Lens' BatchCreateIndex (Maybe ObjectReference)

A reference to the parent object that contains the index object.

batchCreateIndex_linkName :: Lens' BatchCreateIndex (Maybe Text)

The name of the link between the parent object and the index object.

batchCreateIndex_batchReferenceName :: Lens' BatchCreateIndex (Maybe Text)

The batch reference name. See Transaction Support for more information.

batchCreateIndex_orderedIndexedAttributeList :: Lens' BatchCreateIndex [AttributeKey]

Specifies the attributes that should be indexed on. Currently only a single attribute is supported.

batchCreateIndex_isUnique :: Lens' BatchCreateIndex Bool

Indicates whether the attribute that is being indexed has unique values or not.