Network.AWS.CloudDirectory.Types.BatchDetachPolicy

Description

Documentation

data BatchDetachPolicy

Detaches the specified policy from the specified directory inside a BatchWrite operation. For more information, see DetachPolicy and BatchWriteRequest$Operations.

See: newBatchDetachPolicy smart constructor.

Constructors

BatchDetachPolicy'

Fields

policyReference :: ObjectReference

Reference that identifies the policy object.

objectReference :: ObjectReference

Reference that identifies the object whose policy object will be detached.

newBatchDetachPolicy

Arguments

:: ObjectReference
$sel:policyReference:BatchDetachPolicy'

-> ObjectReference
$sel:objectReference:BatchDetachPolicy'

-> BatchDetachPolicy

Create a value of BatchDetachPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyReference:BatchDetachPolicy', batchDetachPolicy_policyReference - Reference that identifies the policy object.

$sel:objectReference:BatchDetachPolicy', batchDetachPolicy_objectReference - Reference that identifies the object whose policy object will be detached.

batchDetachPolicy_policyReference :: Lens' BatchDetachPolicy ObjectReference

Reference that identifies the policy object.

batchDetachPolicy_objectReference :: Lens' BatchDetachPolicy ObjectReference

Reference that identifies the object whose policy object will be detached.