Network.AWS.CloudDirectory.Types.BatchListIncomingTypedLinks

Description

Documentation

data BatchListIncomingTypedLinks

Returns a paginated list of all the incoming TypedLinkSpecifier information for an object inside a BatchRead operation. For more information, see ListIncomingTypedLinks and BatchReadRequest$Operations.

See: newBatchListIncomingTypedLinks smart constructor.

Constructors

BatchListIncomingTypedLinks'

Fields

nextToken :: Maybe Text

The pagination token.

filterTypedLink :: Maybe TypedLinkSchemaAndFacetName

Filters are interpreted in the order of the attributes on the typed link facet, not the order in which they are supplied to any API calls.

maxResults :: Maybe Natural

The maximum number of results to retrieve.

filterAttributeRanges :: Maybe [TypedLinkAttributeRange]

Provides range filters for multiple attributes. When providing ranges to typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range.

objectReference :: ObjectReference

The reference that identifies the object whose attributes will be listed.

newBatchListIncomingTypedLinks

Arguments

:: ObjectReference
$sel:objectReference:BatchListIncomingTypedLinks'

-> BatchListIncomingTypedLinks

Create a value of BatchListIncomingTypedLinks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:BatchListIncomingTypedLinks', batchListIncomingTypedLinks_nextToken - The pagination token.

$sel:filterTypedLink:BatchListIncomingTypedLinks', batchListIncomingTypedLinks_filterTypedLink - Filters are interpreted in the order of the attributes on the typed link facet, not the order in which they are supplied to any API calls.

$sel:maxResults:BatchListIncomingTypedLinks', batchListIncomingTypedLinks_maxResults - The maximum number of results to retrieve.

$sel:filterAttributeRanges:BatchListIncomingTypedLinks', batchListIncomingTypedLinks_filterAttributeRanges - Provides range filters for multiple attributes. When providing ranges to typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range.

$sel:objectReference:BatchListIncomingTypedLinks', batchListIncomingTypedLinks_objectReference - The reference that identifies the object whose attributes will be listed.

batchListIncomingTypedLinks_nextToken :: Lens' BatchListIncomingTypedLinks (Maybe Text)

The pagination token.

batchListIncomingTypedLinks_filterTypedLink :: Lens' BatchListIncomingTypedLinks (Maybe TypedLinkSchemaAndFacetName)

Filters are interpreted in the order of the attributes on the typed link facet, not the order in which they are supplied to any API calls.

batchListIncomingTypedLinks_maxResults :: Lens' BatchListIncomingTypedLinks (Maybe Natural)

The maximum number of results to retrieve.

batchListIncomingTypedLinks_filterAttributeRanges :: Lens' BatchListIncomingTypedLinks (Maybe [TypedLinkAttributeRange])

Provides range filters for multiple attributes. When providing ranges to typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range.

batchListIncomingTypedLinks_objectReference :: Lens' BatchListIncomingTypedLinks ObjectReference

The reference that identifies the object whose attributes will be listed.