Network.AWS.CloudDirectory.Types.BatchListIncomingTypedLinksResponse

Description

Documentation

data BatchListIncomingTypedLinksResponse

Represents the output of a ListIncomingTypedLinks response operation.

See: newBatchListIncomingTypedLinksResponse smart constructor.

Constructors

BatchListIncomingTypedLinksResponse'

Fields

linkSpecifiers :: Maybe [TypedLinkSpecifier]

Returns one or more typed link specifiers as output.

nextToken :: Maybe Text

The pagination token.

newBatchListIncomingTypedLinksResponse :: BatchListIncomingTypedLinksResponse

Create a value of BatchListIncomingTypedLinksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:linkSpecifiers:BatchListIncomingTypedLinksResponse', batchListIncomingTypedLinksResponse_linkSpecifiers - Returns one or more typed link specifiers as output.

$sel:nextToken:BatchListIncomingTypedLinksResponse', batchListIncomingTypedLinksResponse_nextToken - The pagination token.

batchListIncomingTypedLinksResponse_linkSpecifiers :: Lens' BatchListIncomingTypedLinksResponse (Maybe [TypedLinkSpecifier])

Returns one or more typed link specifiers as output.

batchListIncomingTypedLinksResponse_nextToken :: Lens' BatchListIncomingTypedLinksResponse (Maybe Text)

The pagination token.