Network.AWS.CloudDirectory.Types.BatchListIndex

Description

Documentation

data BatchListIndex

Lists objects attached to the specified index inside a BatchRead operation. For more information, see ListIndex and BatchReadRequest$Operations.

See: newBatchListIndex smart constructor.

Constructors

BatchListIndex'

Fields

rangesOnIndexedValues :: Maybe [ObjectAttributeRange]

Specifies the ranges of indexed values that you want to query.

nextToken :: Maybe Text

The pagination token.

maxResults :: Maybe Natural

The maximum number of results to retrieve.

indexReference :: ObjectReference

The reference to the index to list.

newBatchListIndex

Arguments

:: ObjectReference
$sel:indexReference:BatchListIndex'

-> BatchListIndex

Create a value of BatchListIndex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rangesOnIndexedValues:BatchListIndex', batchListIndex_rangesOnIndexedValues - Specifies the ranges of indexed values that you want to query.

$sel:nextToken:BatchListIndex', batchListIndex_nextToken - The pagination token.

$sel:maxResults:BatchListIndex', batchListIndex_maxResults - The maximum number of results to retrieve.

$sel:indexReference:BatchListIndex', batchListIndex_indexReference - The reference to the index to list.

batchListIndex_rangesOnIndexedValues :: Lens' BatchListIndex (Maybe [ObjectAttributeRange])

Specifies the ranges of indexed values that you want to query.

batchListIndex_nextToken :: Lens' BatchListIndex (Maybe Text)

The pagination token.

batchListIndex_maxResults :: Lens' BatchListIndex (Maybe Natural)

The maximum number of results to retrieve.

batchListIndex_indexReference :: Lens' BatchListIndex ObjectReference

The reference to the index to list.