Network.AWS.CloudDirectory.Types.BatchReadOperation

Description

Documentation

data BatchReadOperation

Represents the output of a BatchRead operation.

See: newBatchReadOperation smart constructor.

Constructors

BatchReadOperation'

Fields

listIndex :: Maybe BatchListIndex

Lists objects attached to the specified index.

getObjectInformation :: Maybe BatchGetObjectInformation

Retrieves metadata about an object.

listAttachedIndices :: Maybe BatchListAttachedIndices

Lists indices attached to an object.

lookupPolicy :: Maybe BatchLookupPolicy

Lists all policies from the root of the Directory to the object specified. If there are no policies present, an empty list is returned. If policies are present, and if some objects don't have the policies attached, it returns the ObjectIdentifier for such objects. If policies are present, it returns ObjectIdentifier, policyId, and policyType. Paths that don't lead to the root from the target object are ignored. For more information, see Policies.

listObjectParentPaths :: Maybe BatchListObjectParentPaths

Retrieves all available parent paths for any object type such as node, leaf node, policy node, and index node objects. For more information about objects, see Directory Structure.

listObjectAttributes :: Maybe BatchListObjectAttributes

Lists all attributes that are associated with an object.

listIncomingTypedLinks :: Maybe BatchListIncomingTypedLinks

Returns a paginated list of all the incoming TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

getLinkAttributes :: Maybe BatchGetLinkAttributes

Retrieves attributes that are associated with a typed link.

getObjectAttributes :: Maybe BatchGetObjectAttributes

Retrieves attributes within a facet that are associated with an object.

listObjectChildren :: Maybe BatchListObjectChildren

Returns a paginated list of child objects that are associated with a given object.

listObjectParents :: Maybe BatchListObjectParents

Lists parent objects that are associated with a given object in pagination fashion.

listPolicyAttachments :: Maybe BatchListPolicyAttachments

Returns all of the ObjectIdentifiers to which a given policy is attached.

listOutgoingTypedLinks :: Maybe BatchListOutgoingTypedLinks

Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

listObjectPolicies :: Maybe BatchListObjectPolicies

Returns policies attached to an object in pagination fashion.

newBatchReadOperation :: BatchReadOperation

Create a value of BatchReadOperation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:listIndex:BatchReadOperation', batchReadOperation_listIndex - Lists objects attached to the specified index.

$sel:getObjectInformation:BatchReadOperation', batchReadOperation_getObjectInformation - Retrieves metadata about an object.

$sel:listAttachedIndices:BatchReadOperation', batchReadOperation_listAttachedIndices - Lists indices attached to an object.

$sel:lookupPolicy:BatchReadOperation', batchReadOperation_lookupPolicy - Lists all policies from the root of the Directory to the object specified. If there are no policies present, an empty list is returned. If policies are present, and if some objects don't have the policies attached, it returns the ObjectIdentifier for such objects. If policies are present, it returns ObjectIdentifier, policyId, and policyType. Paths that don't lead to the root from the target object are ignored. For more information, see Policies.

$sel:listObjectParentPaths:BatchReadOperation', batchReadOperation_listObjectParentPaths - Retrieves all available parent paths for any object type such as node, leaf node, policy node, and index node objects. For more information about objects, see Directory Structure.

$sel:listObjectAttributes:BatchReadOperation', batchReadOperation_listObjectAttributes - Lists all attributes that are associated with an object.

$sel:listIncomingTypedLinks:BatchReadOperation', batchReadOperation_listIncomingTypedLinks - Returns a paginated list of all the incoming TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

$sel:getLinkAttributes:BatchReadOperation', batchReadOperation_getLinkAttributes - Retrieves attributes that are associated with a typed link.

$sel:getObjectAttributes:BatchReadOperation', batchReadOperation_getObjectAttributes - Retrieves attributes within a facet that are associated with an object.

$sel:listObjectChildren:BatchReadOperation', batchReadOperation_listObjectChildren - Returns a paginated list of child objects that are associated with a given object.

$sel:listObjectParents:BatchReadOperation', batchReadOperation_listObjectParents - Lists parent objects that are associated with a given object in pagination fashion.

$sel:listPolicyAttachments:BatchReadOperation', batchReadOperation_listPolicyAttachments - Returns all of the ObjectIdentifiers to which a given policy is attached.

$sel:listOutgoingTypedLinks:BatchReadOperation', batchReadOperation_listOutgoingTypedLinks - Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

$sel:listObjectPolicies:BatchReadOperation', batchReadOperation_listObjectPolicies - Returns policies attached to an object in pagination fashion.

batchReadOperation_listIndex :: Lens' BatchReadOperation (Maybe BatchListIndex)

Lists objects attached to the specified index.

batchReadOperation_getObjectInformation :: Lens' BatchReadOperation (Maybe BatchGetObjectInformation)

Retrieves metadata about an object.

batchReadOperation_listAttachedIndices :: Lens' BatchReadOperation (Maybe BatchListAttachedIndices)

Lists indices attached to an object.

batchReadOperation_lookupPolicy :: Lens' BatchReadOperation (Maybe BatchLookupPolicy)

Lists all policies from the root of the Directory to the object specified. If there are no policies present, an empty list is returned. If policies are present, and if some objects don't have the policies attached, it returns the ObjectIdentifier for such objects. If policies are present, it returns ObjectIdentifier, policyId, and policyType. Paths that don't lead to the root from the target object are ignored. For more information, see Policies.

batchReadOperation_listObjectParentPaths :: Lens' BatchReadOperation (Maybe BatchListObjectParentPaths)

Retrieves all available parent paths for any object type such as node, leaf node, policy node, and index node objects. For more information about objects, see Directory Structure.

batchReadOperation_listObjectAttributes :: Lens' BatchReadOperation (Maybe BatchListObjectAttributes)

Lists all attributes that are associated with an object.

batchReadOperation_listIncomingTypedLinks :: Lens' BatchReadOperation (Maybe BatchListIncomingTypedLinks)

Returns a paginated list of all the incoming TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

batchReadOperation_getLinkAttributes :: Lens' BatchReadOperation (Maybe BatchGetLinkAttributes)

Retrieves attributes that are associated with a typed link.

batchReadOperation_getObjectAttributes :: Lens' BatchReadOperation (Maybe BatchGetObjectAttributes)

Retrieves attributes within a facet that are associated with an object.

batchReadOperation_listObjectChildren :: Lens' BatchReadOperation (Maybe BatchListObjectChildren)

Returns a paginated list of child objects that are associated with a given object.

batchReadOperation_listObjectParents :: Lens' BatchReadOperation (Maybe BatchListObjectParents)

Lists parent objects that are associated with a given object in pagination fashion.

batchReadOperation_listPolicyAttachments :: Lens' BatchReadOperation (Maybe BatchListPolicyAttachments)

Returns all of the ObjectIdentifiers to which a given policy is attached.

batchReadOperation_listOutgoingTypedLinks :: Lens' BatchReadOperation (Maybe BatchListOutgoingTypedLinks)

Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

batchReadOperation_listObjectPolicies :: Lens' BatchReadOperation (Maybe BatchListObjectPolicies)

Returns policies attached to an object in pagination fashion.