Network.AWS.CloudDirectory.Types.BatchRemoveFacetFromObject

Description

Documentation

data BatchRemoveFacetFromObject

A batch operation to remove a facet from an object.

See: newBatchRemoveFacetFromObject smart constructor.

Constructors

BatchRemoveFacetFromObject'

Fields

schemaFacet :: SchemaFacet

The facet to remove from the object.

objectReference :: ObjectReference

A reference to the object whose facet will be removed.

newBatchRemoveFacetFromObject

Arguments

:: SchemaFacet
$sel:schemaFacet:BatchRemoveFacetFromObject'

-> ObjectReference
$sel:objectReference:BatchRemoveFacetFromObject'

-> BatchRemoveFacetFromObject

Create a value of BatchRemoveFacetFromObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaFacet:BatchRemoveFacetFromObject', batchRemoveFacetFromObject_schemaFacet - The facet to remove from the object.

$sel:objectReference:BatchRemoveFacetFromObject', batchRemoveFacetFromObject_objectReference - A reference to the object whose facet will be removed.

batchRemoveFacetFromObject_schemaFacet :: Lens' BatchRemoveFacetFromObject SchemaFacet

The facet to remove from the object.

batchRemoveFacetFromObject_objectReference :: Lens' BatchRemoveFacetFromObject ObjectReference

A reference to the object whose facet will be removed.