Network.AWS.CloudDirectory.Types.BatchWriteOperation

Description

Documentation

data BatchWriteOperation

Represents the output of a BatchWrite operation.

See: newBatchWriteOperation smart constructor.

Constructors

BatchWriteOperation'

Fields

attachTypedLink :: Maybe BatchAttachTypedLink

Attaches a typed link to a specified source and target object. For more information, see Typed Links.

deleteObject :: Maybe BatchDeleteObject

Deletes an object in a Directory.

createObject :: Maybe BatchCreateObject

Creates an object.

updateLinkAttributes :: Maybe BatchUpdateLinkAttributes

Updates a given object's attributes.

detachTypedLink :: Maybe BatchDetachTypedLink

Detaches a typed link from a specified source and target object. For more information, see Typed Links.

createIndex :: Maybe BatchCreateIndex

Creates an index object. See Indexing and search for more information.

detachFromIndex :: Maybe BatchDetachFromIndex

Detaches the specified object from the specified index.

detachPolicy :: Maybe BatchDetachPolicy

Detaches a policy from a Directory.

attachObject :: Maybe BatchAttachObject

Attaches an object to a Directory.

attachToIndex :: Maybe BatchAttachToIndex

Attaches the specified object to the specified index.

attachPolicy :: Maybe BatchAttachPolicy

Attaches a policy object to a regular object. An object can have a limited number of attached policies.

updateObjectAttributes :: Maybe BatchUpdateObjectAttributes

Updates a given object's attributes.

removeFacetFromObject :: Maybe BatchRemoveFacetFromObject

A batch operation that removes a facet from an object.

detachObject :: Maybe BatchDetachObject

Detaches an object from a Directory.

addFacetToObject :: Maybe BatchAddFacetToObject

A batch operation that adds a facet to an object.

newBatchWriteOperation :: BatchWriteOperation

Create a value of BatchWriteOperation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attachTypedLink:BatchWriteOperation', batchWriteOperation_attachTypedLink - Attaches a typed link to a specified source and target object. For more information, see Typed Links.

$sel:deleteObject:BatchWriteOperation', batchWriteOperation_deleteObject - Deletes an object in a Directory.

$sel:createObject:BatchWriteOperation', batchWriteOperation_createObject - Creates an object.

$sel:updateLinkAttributes:BatchWriteOperation', batchWriteOperation_updateLinkAttributes - Updates a given object's attributes.

$sel:detachTypedLink:BatchWriteOperation', batchWriteOperation_detachTypedLink - Detaches a typed link from a specified source and target object. For more information, see Typed Links.

$sel:createIndex:BatchWriteOperation', batchWriteOperation_createIndex - Creates an index object. See Indexing and search for more information.

$sel:detachFromIndex:BatchWriteOperation', batchWriteOperation_detachFromIndex - Detaches the specified object from the specified index.

$sel:detachPolicy:BatchWriteOperation', batchWriteOperation_detachPolicy - Detaches a policy from a Directory.

$sel:attachObject:BatchWriteOperation', batchWriteOperation_attachObject - Attaches an object to a Directory.

$sel:attachToIndex:BatchWriteOperation', batchWriteOperation_attachToIndex - Attaches the specified object to the specified index.

$sel:attachPolicy:BatchWriteOperation', batchWriteOperation_attachPolicy - Attaches a policy object to a regular object. An object can have a limited number of attached policies.

$sel:updateObjectAttributes:BatchWriteOperation', batchWriteOperation_updateObjectAttributes - Updates a given object's attributes.

$sel:removeFacetFromObject:BatchWriteOperation', batchWriteOperation_removeFacetFromObject - A batch operation that removes a facet from an object.

$sel:detachObject:BatchWriteOperation', batchWriteOperation_detachObject - Detaches an object from a Directory.

$sel:addFacetToObject:BatchWriteOperation', batchWriteOperation_addFacetToObject - A batch operation that adds a facet to an object.

batchWriteOperation_attachTypedLink :: Lens' BatchWriteOperation (Maybe BatchAttachTypedLink)

Attaches a typed link to a specified source and target object. For more information, see Typed Links.

batchWriteOperation_deleteObject :: Lens' BatchWriteOperation (Maybe BatchDeleteObject)

Deletes an object in a Directory.

batchWriteOperation_createObject :: Lens' BatchWriteOperation (Maybe BatchCreateObject)

Creates an object.

batchWriteOperation_updateLinkAttributes :: Lens' BatchWriteOperation (Maybe BatchUpdateLinkAttributes)

Updates a given object's attributes.

batchWriteOperation_detachTypedLink :: Lens' BatchWriteOperation (Maybe BatchDetachTypedLink)

Detaches a typed link from a specified source and target object. For more information, see Typed Links.

batchWriteOperation_createIndex :: Lens' BatchWriteOperation (Maybe BatchCreateIndex)

Creates an index object. See Indexing and search for more information.

batchWriteOperation_detachFromIndex :: Lens' BatchWriteOperation (Maybe BatchDetachFromIndex)

Detaches the specified object from the specified index.

batchWriteOperation_detachPolicy :: Lens' BatchWriteOperation (Maybe BatchDetachPolicy)

Detaches a policy from a Directory.

batchWriteOperation_attachObject :: Lens' BatchWriteOperation (Maybe BatchAttachObject)

Attaches an object to a Directory.

batchWriteOperation_attachToIndex :: Lens' BatchWriteOperation (Maybe BatchAttachToIndex)

Attaches the specified object to the specified index.

batchWriteOperation_attachPolicy :: Lens' BatchWriteOperation (Maybe BatchAttachPolicy)

Attaches a policy object to a regular object. An object can have a limited number of attached policies.

batchWriteOperation_updateObjectAttributes :: Lens' BatchWriteOperation (Maybe BatchUpdateObjectAttributes)

Updates a given object's attributes.

batchWriteOperation_removeFacetFromObject :: Lens' BatchWriteOperation (Maybe BatchRemoveFacetFromObject)

A batch operation that removes a facet from an object.

batchWriteOperation_detachObject :: Lens' BatchWriteOperation (Maybe BatchDetachObject)

Detaches an object from a Directory.

batchWriteOperation_addFacetToObject :: Lens' BatchWriteOperation (Maybe BatchAddFacetToObject)

A batch operation that adds a facet to an object.