Network.AWS.CloudDirectory.Types.IndexAttachment

Description

Documentation

data IndexAttachment

Represents an index and an attached object.

See: newIndexAttachment smart constructor.

Constructors

IndexAttachment'

Fields

indexedAttributes :: Maybe [AttributeKeyAndValue]

The indexed attribute values.

objectIdentifier :: Maybe Text

In response to ListIndex, the ObjectIdentifier of the object attached to the index. In response to ListAttachedIndices, the ObjectIdentifier of the index attached to the object. This field will always contain the ObjectIdentifier of the object on the opposite side of the attachment specified in the query.

newIndexAttachment :: IndexAttachment

Create a value of IndexAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:indexedAttributes:IndexAttachment', indexAttachment_indexedAttributes - The indexed attribute values.

$sel:objectIdentifier:IndexAttachment', indexAttachment_objectIdentifier - In response to ListIndex, the ObjectIdentifier of the object attached to the index. In response to ListAttachedIndices, the ObjectIdentifier of the index attached to the object. This field will always contain the ObjectIdentifier of the object on the opposite side of the attachment specified in the query.

indexAttachment_indexedAttributes :: Lens' IndexAttachment (Maybe [AttributeKeyAndValue])

The indexed attribute values.

indexAttachment_objectIdentifier :: Lens' IndexAttachment (Maybe Text)

In response to ListIndex, the ObjectIdentifier of the object attached to the index. In response to ListAttachedIndices, the ObjectIdentifier of the index attached to the object. This field will always contain the ObjectIdentifier of the object on the opposite side of the attachment specified in the query.