Network.AWS.CloudDirectory.Types.TypedLinkAttributeRange

Description

Documentation

data TypedLinkAttributeRange

Identifies the range of attributes that are used by a specified filter.

See: newTypedLinkAttributeRange smart constructor.

Constructors

TypedLinkAttributeRange'

Fields

attributeName :: Maybe Text

The unique name of the typed link attribute.

range :: TypedAttributeValueRange

The range of attribute values that are being selected.

newTypedLinkAttributeRange

Arguments

:: TypedAttributeValueRange
$sel:range:TypedLinkAttributeRange'

-> TypedLinkAttributeRange

Create a value of TypedLinkAttributeRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeName:TypedLinkAttributeRange', typedLinkAttributeRange_attributeName - The unique name of the typed link attribute.

$sel:range:TypedLinkAttributeRange', typedLinkAttributeRange_range - The range of attribute values that are being selected.

typedLinkAttributeRange_attributeName :: Lens' TypedLinkAttributeRange (Maybe Text)

The unique name of the typed link attribute.

typedLinkAttributeRange_range :: Lens' TypedLinkAttributeRange TypedAttributeValueRange

The range of attribute values that are being selected.