Network.AWS.CloudDirectory.UpdateFacet

Description

Does the following:

Adds new Attributes, Rules, or ObjectTypes.

Updates existing Attributes, Rules, or ObjectTypes.

Deletes existing Attributes, Rules, or ObjectTypes.

Creating a Request

data UpdateFacet

See: newUpdateFacet smart constructor.

Constructors

UpdateFacet'

Fields

attributeUpdates :: Maybe [FacetAttributeUpdate]

List of attributes that need to be updated in a given schema Facet. Each attribute is followed by AttributeAction, which specifies the type of update operation to perform.

objectType :: Maybe ObjectType

The object type that is associated with the facet. See CreateFacetRequest$ObjectType for more details.

schemaArn :: Text

The Amazon Resource Name (ARN) that is associated with the Facet. For more information, see arns.

name :: Text

The name of the facet.

newUpdateFacet

Arguments

:: Text
$sel:schemaArn:UpdateFacet'

-> Text
$sel:name:UpdateFacet'

-> UpdateFacet

Create a value of UpdateFacet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeUpdates:UpdateFacet', updateFacet_attributeUpdates - List of attributes that need to be updated in a given schema Facet. Each attribute is followed by AttributeAction, which specifies the type of update operation to perform.

$sel:objectType:UpdateFacet', updateFacet_objectType - The object type that is associated with the facet. See CreateFacetRequest$ObjectType for more details.

$sel:schemaArn:UpdateFacet', updateFacet_schemaArn - The Amazon Resource Name (ARN) that is associated with the Facet. For more information, see arns.

$sel:name:UpdateFacet', updateFacet_name - The name of the facet.

Request Lenses

updateFacet_attributeUpdates :: Lens' UpdateFacet (Maybe [FacetAttributeUpdate])

List of attributes that need to be updated in a given schema Facet. Each attribute is followed by AttributeAction, which specifies the type of update operation to perform.

updateFacet_objectType :: Lens' UpdateFacet (Maybe ObjectType)

The object type that is associated with the facet. See CreateFacetRequest$ObjectType for more details.

updateFacet_schemaArn :: Lens' UpdateFacet Text

The Amazon Resource Name (ARN) that is associated with the Facet. For more information, see arns.

updateFacet_name :: Lens' UpdateFacet Text

The name of the facet.

Destructuring the Response

data UpdateFacetResponse

See: newUpdateFacetResponse smart constructor.

Constructors

UpdateFacetResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateFacetResponse

Arguments

:: Int
$sel:httpStatus:UpdateFacetResponse'

-> UpdateFacetResponse

Create a value of UpdateFacetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateFacetResponse', updateFacetResponse_httpStatus - The response's http status code.

Response Lenses

updateFacetResponse_httpStatus :: Lens' UpdateFacetResponse Int

The response's http status code.