Network.AWS.CloudDirectory.UpdateLinkAttributes

Description

Updates a given typed link’s attributes. Attributes to be updated must not contribute to the typed link’s identity, as defined by its IdentityAttributeOrder.

Creating a Request

data UpdateLinkAttributes

See: newUpdateLinkAttributes smart constructor.

Constructors

UpdateLinkAttributes'

Fields

directoryArn :: Text

The Amazon Resource Name (ARN) that is associated with the Directory where the updated typed link resides. For more information, see arns or Typed Links.

typedLinkSpecifier :: TypedLinkSpecifier

Allows a typed link specifier to be accepted as input.

attributeUpdates :: [LinkAttributeUpdate]

The attributes update structure.

newUpdateLinkAttributes

Arguments

:: Text
$sel:directoryArn:UpdateLinkAttributes'

-> TypedLinkSpecifier
$sel:typedLinkSpecifier:UpdateLinkAttributes'

-> UpdateLinkAttributes

Create a value of UpdateLinkAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryArn:UpdateLinkAttributes', updateLinkAttributes_directoryArn - The Amazon Resource Name (ARN) that is associated with the Directory where the updated typed link resides. For more information, see arns or Typed Links.

$sel:typedLinkSpecifier:UpdateLinkAttributes', updateLinkAttributes_typedLinkSpecifier - Allows a typed link specifier to be accepted as input.

$sel:attributeUpdates:UpdateLinkAttributes', updateLinkAttributes_attributeUpdates - The attributes update structure.

Request Lenses

updateLinkAttributes_directoryArn :: Lens' UpdateLinkAttributes Text

The Amazon Resource Name (ARN) that is associated with the Directory where the updated typed link resides. For more information, see arns or Typed Links.

updateLinkAttributes_typedLinkSpecifier :: Lens' UpdateLinkAttributes TypedLinkSpecifier

Allows a typed link specifier to be accepted as input.

updateLinkAttributes_attributeUpdates :: Lens' UpdateLinkAttributes [LinkAttributeUpdate]

The attributes update structure.

Destructuring the Response

data UpdateLinkAttributesResponse

See: newUpdateLinkAttributesResponse smart constructor.

Constructors

UpdateLinkAttributesResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateLinkAttributesResponse

Arguments

:: Int
$sel:httpStatus:UpdateLinkAttributesResponse'

-> UpdateLinkAttributesResponse

Create a value of UpdateLinkAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateLinkAttributesResponse', updateLinkAttributesResponse_httpStatus - The response's http status code.

Response Lenses

updateLinkAttributesResponse_httpStatus :: Lens' UpdateLinkAttributesResponse Int

The response's http status code.