Network.AWS.CloudDirectory.UpdateObjectAttributes

Description

Updates a given object's attributes.

Creating a Request

data UpdateObjectAttributes

See: newUpdateObjectAttributes smart constructor.

Constructors

UpdateObjectAttributes'

Fields

directoryArn :: Text

The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns.

objectReference :: ObjectReference

The reference that identifies the object.

attributeUpdates :: [ObjectAttributeUpdate]

The attributes update structure.

newUpdateObjectAttributes

Arguments

:: Text
$sel:directoryArn:UpdateObjectAttributes'

-> ObjectReference
$sel:objectReference:UpdateObjectAttributes'

-> UpdateObjectAttributes

Create a value of UpdateObjectAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryArn:UpdateObjectAttributes', updateObjectAttributes_directoryArn - The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns.

$sel:objectReference:UpdateObjectAttributes', updateObjectAttributes_objectReference - The reference that identifies the object.

$sel:attributeUpdates:UpdateObjectAttributes', updateObjectAttributes_attributeUpdates - The attributes update structure.

Request Lenses

updateObjectAttributes_directoryArn :: Lens' UpdateObjectAttributes Text

The Amazon Resource Name (ARN) that is associated with the Directory where the object resides. For more information, see arns.

updateObjectAttributes_objectReference :: Lens' UpdateObjectAttributes ObjectReference

The reference that identifies the object.

updateObjectAttributes_attributeUpdates :: Lens' UpdateObjectAttributes [ObjectAttributeUpdate]

The attributes update structure.

Destructuring the Response

data UpdateObjectAttributesResponse

See: newUpdateObjectAttributesResponse smart constructor.

Constructors

UpdateObjectAttributesResponse'

Fields

objectIdentifier :: Maybe Text

The ObjectIdentifier of the updated object.

httpStatus :: Int

The response's http status code.

newUpdateObjectAttributesResponse

Arguments

:: Int
$sel:httpStatus:UpdateObjectAttributesResponse'

-> UpdateObjectAttributesResponse

Create a value of UpdateObjectAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:objectIdentifier:UpdateObjectAttributesResponse', updateObjectAttributesResponse_objectIdentifier - The ObjectIdentifier of the updated object.

$sel:httpStatus:UpdateObjectAttributesResponse', updateObjectAttributesResponse_httpStatus - The response's http status code.

Response Lenses

updateObjectAttributesResponse_objectIdentifier :: Lens' UpdateObjectAttributesResponse (Maybe Text)

The ObjectIdentifier of the updated object.

updateObjectAttributesResponse_httpStatus :: Lens' UpdateObjectAttributesResponse Int

The response's http status code.