Network.AWS.CloudDirectory.UpdateTypedLinkFacet

Description

Updates a TypedLinkFacet. For more information, see Typed Links.

Creating a Request

data UpdateTypedLinkFacet

See: newUpdateTypedLinkFacet smart constructor.

Constructors

UpdateTypedLinkFacet'

Fields

schemaArn :: Text

The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns.

name :: Text

The unique name of the typed link facet.

attributeUpdates :: [TypedLinkFacetAttributeUpdate]

Attributes update structure.

identityAttributeOrder :: [Text]

The order of identity attributes for the facet, from most significant to least significant. The ability to filter typed links considers the order that the attributes are defined on the typed link facet. When providing ranges to a typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range. Filters are interpreted in the order of the attributes on the typed link facet, not the order in which they are supplied to any API calls. For more information about identity attributes, see Typed Links.

newUpdateTypedLinkFacet

Arguments

:: Text
$sel:schemaArn:UpdateTypedLinkFacet'

-> Text
$sel:name:UpdateTypedLinkFacet'

-> UpdateTypedLinkFacet

Create a value of UpdateTypedLinkFacet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaArn:UpdateTypedLinkFacet', updateTypedLinkFacet_schemaArn - The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns.

$sel:name:UpdateTypedLinkFacet', updateTypedLinkFacet_name - The unique name of the typed link facet.

$sel:attributeUpdates:UpdateTypedLinkFacet', updateTypedLinkFacet_attributeUpdates - Attributes update structure.

$sel:identityAttributeOrder:UpdateTypedLinkFacet', updateTypedLinkFacet_identityAttributeOrder - The order of identity attributes for the facet, from most significant to least significant. The ability to filter typed links considers the order that the attributes are defined on the typed link facet. When providing ranges to a typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range. Filters are interpreted in the order of the attributes on the typed link facet, not the order in which they are supplied to any API calls. For more information about identity attributes, see Typed Links.

Request Lenses

updateTypedLinkFacet_schemaArn :: Lens' UpdateTypedLinkFacet Text

The Amazon Resource Name (ARN) that is associated with the schema. For more information, see arns.

updateTypedLinkFacet_name :: Lens' UpdateTypedLinkFacet Text

The unique name of the typed link facet.

updateTypedLinkFacet_attributeUpdates :: Lens' UpdateTypedLinkFacet [TypedLinkFacetAttributeUpdate]

Attributes update structure.

updateTypedLinkFacet_identityAttributeOrder :: Lens' UpdateTypedLinkFacet [Text]

The order of identity attributes for the facet, from most significant to least significant. The ability to filter typed links considers the order that the attributes are defined on the typed link facet. When providing ranges to a typed link selection, any inexact ranges must be specified at the end. Any attributes that do not have a range specified are presumed to match the entire range. Filters are interpreted in the order of the attributes on the typed link facet, not the order in which they are supplied to any API calls. For more information about identity attributes, see Typed Links.

Destructuring the Response

data UpdateTypedLinkFacetResponse

See: newUpdateTypedLinkFacetResponse smart constructor.

Constructors

UpdateTypedLinkFacetResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateTypedLinkFacetResponse

Arguments

:: Int
$sel:httpStatus:UpdateTypedLinkFacetResponse'

-> UpdateTypedLinkFacetResponse

Create a value of UpdateTypedLinkFacetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateTypedLinkFacetResponse', updateTypedLinkFacetResponse_httpStatus - The response's http status code.

Response Lenses

updateTypedLinkFacetResponse_httpStatus :: Lens' UpdateTypedLinkFacetResponse Int

The response's http status code.