Network.AWS.CloudDirectory.UpgradeAppliedSchema

Description

Upgrades a single directory in-place using the PublishedSchemaArn with schema updates found in MinorVersion. Backwards-compatible minor version upgrades are instantaneously available for readers on all objects in the directory. Note: This is a synchronous API call and upgrades only one schema on a given directory per call. To upgrade multiple directories from one schema, you would need to call this API on each directory.

Creating a Request

data UpgradeAppliedSchema

See: newUpgradeAppliedSchema smart constructor.

Constructors

UpgradeAppliedSchema'

Fields

dryRun :: Maybe Bool

Used for testing whether the major version schemas are backward compatible or not. If schema compatibility fails, an exception would be thrown else the call would succeed but no changes will be saved. This parameter is optional.

publishedSchemaArn :: Text

The revision of the published schema to upgrade the directory to.

directoryArn :: Text

The ARN for the directory to which the upgraded schema will be applied.

newUpgradeAppliedSchema

Arguments

:: Text
$sel:publishedSchemaArn:UpgradeAppliedSchema'

-> Text
$sel:directoryArn:UpgradeAppliedSchema'

-> UpgradeAppliedSchema

Create a value of UpgradeAppliedSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:UpgradeAppliedSchema', upgradeAppliedSchema_dryRun - Used for testing whether the major version schemas are backward compatible or not. If schema compatibility fails, an exception would be thrown else the call would succeed but no changes will be saved. This parameter is optional.

$sel:publishedSchemaArn:UpgradeAppliedSchema', upgradeAppliedSchema_publishedSchemaArn - The revision of the published schema to upgrade the directory to.

$sel:directoryArn:UpgradeAppliedSchema', upgradeAppliedSchema_directoryArn - The ARN for the directory to which the upgraded schema will be applied.

Request Lenses

upgradeAppliedSchema_dryRun :: Lens' UpgradeAppliedSchema (Maybe Bool)

Used for testing whether the major version schemas are backward compatible or not. If schema compatibility fails, an exception would be thrown else the call would succeed but no changes will be saved. This parameter is optional.

upgradeAppliedSchema_publishedSchemaArn :: Lens' UpgradeAppliedSchema Text

The revision of the published schema to upgrade the directory to.

upgradeAppliedSchema_directoryArn :: Lens' UpgradeAppliedSchema Text

The ARN for the directory to which the upgraded schema will be applied.

Destructuring the Response

data UpgradeAppliedSchemaResponse

See: newUpgradeAppliedSchemaResponse smart constructor.

Constructors

UpgradeAppliedSchemaResponse'

Fields

directoryArn :: Maybe Text

The ARN of the directory that is returned as part of the response.

upgradedSchemaArn :: Maybe Text

The ARN of the upgraded schema that is returned as part of the response.

httpStatus :: Int

The response's http status code.

newUpgradeAppliedSchemaResponse

Arguments

:: Int
$sel:httpStatus:UpgradeAppliedSchemaResponse'

-> UpgradeAppliedSchemaResponse

Create a value of UpgradeAppliedSchemaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryArn:UpgradeAppliedSchema', upgradeAppliedSchemaResponse_directoryArn - The ARN of the directory that is returned as part of the response.

$sel:upgradedSchemaArn:UpgradeAppliedSchemaResponse', upgradeAppliedSchemaResponse_upgradedSchemaArn - The ARN of the upgraded schema that is returned as part of the response.

$sel:httpStatus:UpgradeAppliedSchemaResponse', upgradeAppliedSchemaResponse_httpStatus - The response's http status code.

Response Lenses

upgradeAppliedSchemaResponse_directoryArn :: Lens' UpgradeAppliedSchemaResponse (Maybe Text)

The ARN of the directory that is returned as part of the response.

upgradeAppliedSchemaResponse_upgradedSchemaArn :: Lens' UpgradeAppliedSchemaResponse (Maybe Text)

The ARN of the upgraded schema that is returned as part of the response.

upgradeAppliedSchemaResponse_httpStatus :: Lens' UpgradeAppliedSchemaResponse Int

The response's http status code.