Network.AWS.CloudFormation.ActivateType

Description

Activates a public third-party extension, making it available for use in stack templates. For more information, see Using public extensions in the CloudFormation User Guide.

Once you have activated a public third-party extension in your account and region, use SetTypeConfiguration to specify configuration properties for the extension. For more information, see Configuring extensions at the account level in the CloudFormation User Guide.

Creating a Request

data ActivateType

See: newActivateType smart constructor.

Constructors

ActivateType'

Fields

typeName :: Maybe Text

The name of the extension.

Conditional: You must specify PublicTypeArn, or TypeName, Type, and PublisherId.

publicTypeArn :: Maybe Text

The Amazon Resource Number (ARN) of the public extension.

Conditional: You must specify PublicTypeArn, or TypeName, Type, and PublisherId.

loggingConfig :: Maybe LoggingConfig

majorVersion :: Maybe Natural

The major version of this extension you want to activate, if multiple major versions are available. The default is the latest major version. CloudFormation uses the latest available minor version of the major version selected.

You can specify MajorVersion or VersionBump, but not both.

publisherId :: Maybe Text

The ID of the extension publisher.

Conditional: You must specify PublicTypeArn, or TypeName, Type, and PublisherId.

executionRoleArn :: Maybe Text

The name of the IAM execution role to use to activate the extension.

typeNameAlias :: Maybe Text

An alias to assign to the public extension, in this account and region. If you specify an alias for the extension, CloudFormation treats the alias as the extension type name within this account and region. You must use the alias to refer to the extension in your templates, API calls, and CloudFormation console.

An extension alias must be unique within a given account and region. You can activate the same public resource multiple times in the same account and region, using different type name aliases.

autoUpdate :: Maybe Bool

Whether to automatically update the extension in this account and region when a new minor version is published by the extension publisher. Major versions released by the publisher must be manually updated.

The default is true.

versionBump :: Maybe VersionBump

Manually updates a previously-activated type to a new major or minor version, if available. You can also use this parameter to update the value of AutoUpdate.

MAJOR: CloudFormation updates the extension to the newest major version, if one is available.

MINOR: CloudFormation updates the extension to the newest minor version, if one is available.

type' :: Maybe ThirdPartyType

The extension type.

Conditional: You must specify PublicTypeArn, or TypeName, Type, and PublisherId.

newActivateType :: ActivateType

Create a value of ActivateType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:ActivateType', activateType_typeName - The name of the extension.

Conditional: You must specify PublicTypeArn, or TypeName, Type, and PublisherId.

$sel:publicTypeArn:ActivateType', activateType_publicTypeArn - The Amazon Resource Number (ARN) of the public extension.

Conditional: You must specify PublicTypeArn, or TypeName, Type, and PublisherId.

$sel:loggingConfig:ActivateType', activateType_loggingConfig - Undocumented member.

$sel:majorVersion:ActivateType', activateType_majorVersion - The major version of this extension you want to activate, if multiple major versions are available. The default is the latest major version. CloudFormation uses the latest available minor version of the major version selected.

You can specify MajorVersion or VersionBump, but not both.

$sel:publisherId:ActivateType', activateType_publisherId - The ID of the extension publisher.

Conditional: You must specify PublicTypeArn, or TypeName, Type, and PublisherId.

$sel:executionRoleArn:ActivateType', activateType_executionRoleArn - The name of the IAM execution role to use to activate the extension.

$sel:typeNameAlias:ActivateType', activateType_typeNameAlias - An alias to assign to the public extension, in this account and region. If you specify an alias for the extension, CloudFormation treats the alias as the extension type name within this account and region. You must use the alias to refer to the extension in your templates, API calls, and CloudFormation console.

An extension alias must be unique within a given account and region. You can activate the same public resource multiple times in the same account and region, using different type name aliases.

$sel:autoUpdate:ActivateType', activateType_autoUpdate - Whether to automatically update the extension in this account and region when a new minor version is published by the extension publisher. Major versions released by the publisher must be manually updated.

The default is true.

$sel:versionBump:ActivateType', activateType_versionBump - Manually updates a previously-activated type to a new major or minor version, if available. You can also use this parameter to update the value of AutoUpdate.

MAJOR: CloudFormation updates the extension to the newest major version, if one is available.

MINOR: CloudFormation updates the extension to the newest minor version, if one is available.

$sel:type':ActivateType', activateType_type - The extension type.

Conditional: You must specify PublicTypeArn, or TypeName, Type, and PublisherId.

Request Lenses

activateType_typeName :: Lens' ActivateType (Maybe Text)

The name of the extension.

Conditional: You must specify PublicTypeArn, or TypeName, Type, and PublisherId.

activateType_publicTypeArn :: Lens' ActivateType (Maybe Text)

The Amazon Resource Number (ARN) of the public extension.

Conditional: You must specify PublicTypeArn, or TypeName, Type, and PublisherId.

activateType_loggingConfig :: Lens' ActivateType (Maybe LoggingConfig)

Undocumented member.

activateType_majorVersion :: Lens' ActivateType (Maybe Natural)

The major version of this extension you want to activate, if multiple major versions are available. The default is the latest major version. CloudFormation uses the latest available minor version of the major version selected.

You can specify MajorVersion or VersionBump, but not both.

activateType_publisherId :: Lens' ActivateType (Maybe Text)

The ID of the extension publisher.

Conditional: You must specify PublicTypeArn, or TypeName, Type, and PublisherId.

activateType_executionRoleArn :: Lens' ActivateType (Maybe Text)

The name of the IAM execution role to use to activate the extension.

activateType_typeNameAlias :: Lens' ActivateType (Maybe Text)

An alias to assign to the public extension, in this account and region. If you specify an alias for the extension, CloudFormation treats the alias as the extension type name within this account and region. You must use the alias to refer to the extension in your templates, API calls, and CloudFormation console.

An extension alias must be unique within a given account and region. You can activate the same public resource multiple times in the same account and region, using different type name aliases.

activateType_autoUpdate :: Lens' ActivateType (Maybe Bool)

Whether to automatically update the extension in this account and region when a new minor version is published by the extension publisher. Major versions released by the publisher must be manually updated.

The default is true.

activateType_versionBump :: Lens' ActivateType (Maybe VersionBump)

Manually updates a previously-activated type to a new major or minor version, if available. You can also use this parameter to update the value of AutoUpdate.

MAJOR: CloudFormation updates the extension to the newest major version, if one is available.

MINOR: CloudFormation updates the extension to the newest minor version, if one is available.

activateType_type :: Lens' ActivateType (Maybe ThirdPartyType)

The extension type.

Conditional: You must specify PublicTypeArn, or TypeName, Type, and PublisherId.

Destructuring the Response

data ActivateTypeResponse

See: newActivateTypeResponse smart constructor.

Constructors

ActivateTypeResponse'

Fields

arn :: Maybe Text

The Amazon Resource Number (ARN) of the activated extension, in this account and region.

httpStatus :: Int

The response's http status code.

newActivateTypeResponse

Arguments

:: Int
$sel:httpStatus:ActivateTypeResponse'

-> ActivateTypeResponse

Create a value of ActivateTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ActivateTypeResponse', activateTypeResponse_arn - The Amazon Resource Number (ARN) of the activated extension, in this account and region.

$sel:httpStatus:ActivateTypeResponse', activateTypeResponse_httpStatus - The response's http status code.

Response Lenses

activateTypeResponse_arn :: Lens' ActivateTypeResponse (Maybe Text)

The Amazon Resource Number (ARN) of the activated extension, in this account and region.

activateTypeResponse_httpStatus :: Lens' ActivateTypeResponse Int

The response's http status code.