Network.AWS.CloudFormation.CancelUpdateStack

Description

Cancels an update on the specified stack. If the call completes successfully, the stack rolls back the update and reverts to the previous stack configuration.

You can cancel only stacks that are in the UPDATE_IN_PROGRESS state.

Creating a Request

data CancelUpdateStack

The input for the CancelUpdateStack action.

See: newCancelUpdateStack smart constructor.

Constructors

CancelUpdateStack'

Fields

clientRequestToken :: Maybe Text

A unique identifier for this CancelUpdateStack request. Specify this token if you plan to retry requests so that CloudFormation knows that you're not attempting to cancel an update on a stack with the same name. You might retry CancelUpdateStack requests to ensure that CloudFormation successfully received them.

stackName :: Text

The name or the unique stack ID that is associated with the stack.

newCancelUpdateStack

Arguments

:: Text
$sel:stackName:CancelUpdateStack'

-> CancelUpdateStack

Create a value of CancelUpdateStack with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:CancelUpdateStack', cancelUpdateStack_clientRequestToken - A unique identifier for this CancelUpdateStack request. Specify this token if you plan to retry requests so that CloudFormation knows that you're not attempting to cancel an update on a stack with the same name. You might retry CancelUpdateStack requests to ensure that CloudFormation successfully received them.

$sel:stackName:CancelUpdateStack', cancelUpdateStack_stackName - The name or the unique stack ID that is associated with the stack.

Request Lenses

cancelUpdateStack_clientRequestToken :: Lens' CancelUpdateStack (Maybe Text)

A unique identifier for this CancelUpdateStack request. Specify this token if you plan to retry requests so that CloudFormation knows that you're not attempting to cancel an update on a stack with the same name. You might retry CancelUpdateStack requests to ensure that CloudFormation successfully received them.

cancelUpdateStack_stackName :: Lens' CancelUpdateStack Text

The name or the unique stack ID that is associated with the stack.

Destructuring the Response

data CancelUpdateStackResponse

See: newCancelUpdateStackResponse smart constructor.

Constructors

CancelUpdateStackResponse'

newCancelUpdateStackResponse :: CancelUpdateStackResponse

Create a value of CancelUpdateStackResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.