Network.AWS.CloudFormation.DeactivateType

Description

Deactivates a public extension that was previously activated in this account and region.

Once deactivated, an extension cannot be used in any CloudFormation operation. This includes stack update operations where the stack template includes the extension, even if no updates are being made to the extension. In addition, deactivated extensions are not automatically updated if a new version of the extension is released.

Creating a Request

data DeactivateType

See: newDeactivateType smart constructor.

Constructors

DeactivateType'

Fields

typeName :: Maybe Text

The type name of the extension, in this account and region. If you specified a type name alias when enabling the extension, use the type name alias.

Conditional: You must specify either Arn, or TypeName and Type.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the extension, in this account and region.

Conditional: You must specify either Arn, or TypeName and Type.

type' :: Maybe ThirdPartyType

The extension type.

Conditional: You must specify either Arn, or TypeName and Type.

newDeactivateType :: DeactivateType

Create a value of DeactivateType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:DeactivateType', deactivateType_typeName - The type name of the extension, in this account and region. If you specified a type name alias when enabling the extension, use the type name alias.

Conditional: You must specify either Arn, or TypeName and Type.

$sel:arn:DeactivateType', deactivateType_arn - The Amazon Resource Name (ARN) for the extension, in this account and region.

Conditional: You must specify either Arn, or TypeName and Type.

$sel:type':DeactivateType', deactivateType_type - The extension type.

Conditional: You must specify either Arn, or TypeName and Type.

Request Lenses

deactivateType_typeName :: Lens' DeactivateType (Maybe Text)

The type name of the extension, in this account and region. If you specified a type name alias when enabling the extension, use the type name alias.

Conditional: You must specify either Arn, or TypeName and Type.

deactivateType_arn :: Lens' DeactivateType (Maybe Text)

The Amazon Resource Name (ARN) for the extension, in this account and region.

Conditional: You must specify either Arn, or TypeName and Type.

deactivateType_type :: Lens' DeactivateType (Maybe ThirdPartyType)

The extension type.

Conditional: You must specify either Arn, or TypeName and Type.

Destructuring the Response

data DeactivateTypeResponse

See: newDeactivateTypeResponse smart constructor.

Constructors

DeactivateTypeResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeactivateTypeResponse

Arguments

:: Int
$sel:httpStatus:DeactivateTypeResponse'

-> DeactivateTypeResponse

Create a value of DeactivateTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeactivateTypeResponse', deactivateTypeResponse_httpStatus - The response's http status code.

Response Lenses

deactivateTypeResponse_httpStatus :: Lens' DeactivateTypeResponse Int

The response's http status code.