Network.AWS.CloudFormation.DeleteStackInstances

Description

Deletes stack instances for the specified accounts, in the specified Regions.

Creating a Request

data DeleteStackInstances

See: newDeleteStackInstances smart constructor.

Constructors

DeleteStackInstances'

Fields

deploymentTargets :: Maybe DeploymentTargets

Service-managed permissions

The Organizations accounts from which to delete stack instances.

You can specify Accounts or DeploymentTargets, but not both.

operationId :: Maybe Text

The unique identifier for this stack set operation.

If you don't specify an operation ID, the SDK generates one automatically.

The operation ID also functions as an idempotency token, to ensure that CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You can retry stack set operation requests to ensure that CloudFormation successfully received them.

Repeating this stack set operation with a new operation ID retries all stack instances whose status is OUTDATED.

callAs :: Maybe CallAs

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

operationPreferences :: Maybe StackSetOperationPreferences

Preferences for how CloudFormation performs this stack set operation.

accounts :: Maybe [Text]

Self-managed permissions

The names of the Amazon Web Services accounts that you want to delete stack instances for.

You can specify Accounts or DeploymentTargets, but not both.

stackSetName :: Text

The name or unique ID of the stack set that you want to delete stack instances for.

regions :: [Text]

The Regions where you want to delete stack set instances.

retainStacks :: Bool

Removes the stack instances from the specified stack set, but doesn't delete the stacks. You can't reassociate a retained stack or add an existing, saved stack to a new stack set.

For more information, see Stack set operation options.

newDeleteStackInstances

Arguments

:: Text
$sel:stackSetName:DeleteStackInstances'

-> Bool
$sel:retainStacks:DeleteStackInstances'

-> DeleteStackInstances

Create a value of DeleteStackInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentTargets:DeleteStackInstances', deleteStackInstances_deploymentTargets - [Service-managed permissions] The Organizations accounts from which to delete stack instances.

You can specify Accounts or DeploymentTargets, but not both.

$sel:operationId:DeleteStackInstances', deleteStackInstances_operationId - The unique identifier for this stack set operation.

If you don't specify an operation ID, the SDK generates one automatically.

The operation ID also functions as an idempotency token, to ensure that CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You can retry stack set operation requests to ensure that CloudFormation successfully received them.

Repeating this stack set operation with a new operation ID retries all stack instances whose status is OUTDATED.

$sel:callAs:DeleteStackInstances', deleteStackInstances_callAs - [Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

$sel:operationPreferences:DeleteStackInstances', deleteStackInstances_operationPreferences - Preferences for how CloudFormation performs this stack set operation.

$sel:accounts:DeleteStackInstances', deleteStackInstances_accounts - [Self-managed permissions] The names of the Amazon Web Services accounts that you want to delete stack instances for.

You can specify Accounts or DeploymentTargets, but not both.

$sel:stackSetName:DeleteStackInstances', deleteStackInstances_stackSetName - The name or unique ID of the stack set that you want to delete stack instances for.

$sel:regions:DeleteStackInstances', deleteStackInstances_regions - The Regions where you want to delete stack set instances.

$sel:retainStacks:DeleteStackInstances', deleteStackInstances_retainStacks - Removes the stack instances from the specified stack set, but doesn't delete the stacks. You can't reassociate a retained stack or add an existing, saved stack to a new stack set.

For more information, see Stack set operation options.

Request Lenses

deleteStackInstances_deploymentTargets :: Lens' DeleteStackInstances (Maybe DeploymentTargets)

Service-managed permissions

The Organizations accounts from which to delete stack instances.

You can specify Accounts or DeploymentTargets, but not both.

deleteStackInstances_operationId :: Lens' DeleteStackInstances (Maybe Text)

The unique identifier for this stack set operation.

If you don't specify an operation ID, the SDK generates one automatically.

The operation ID also functions as an idempotency token, to ensure that CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You can retry stack set operation requests to ensure that CloudFormation successfully received them.

Repeating this stack set operation with a new operation ID retries all stack instances whose status is OUTDATED.

deleteStackInstances_callAs :: Lens' DeleteStackInstances (Maybe CallAs)

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

deleteStackInstances_operationPreferences :: Lens' DeleteStackInstances (Maybe StackSetOperationPreferences)

Preferences for how CloudFormation performs this stack set operation.

deleteStackInstances_accounts :: Lens' DeleteStackInstances (Maybe [Text])

Self-managed permissions

The names of the Amazon Web Services accounts that you want to delete stack instances for.

You can specify Accounts or DeploymentTargets, but not both.

deleteStackInstances_stackSetName :: Lens' DeleteStackInstances Text

The name or unique ID of the stack set that you want to delete stack instances for.

deleteStackInstances_regions :: Lens' DeleteStackInstances [Text]

The Regions where you want to delete stack set instances.

deleteStackInstances_retainStacks :: Lens' DeleteStackInstances Bool

Removes the stack instances from the specified stack set, but doesn't delete the stacks. You can't reassociate a retained stack or add an existing, saved stack to a new stack set.

For more information, see Stack set operation options.

Destructuring the Response

data DeleteStackInstancesResponse

See: newDeleteStackInstancesResponse smart constructor.

Constructors

DeleteStackInstancesResponse'

Fields

operationId :: Maybe Text

The unique identifier for this stack set operation.

httpStatus :: Int

The response's http status code.

newDeleteStackInstancesResponse

Arguments

:: Int
$sel:httpStatus:DeleteStackInstancesResponse'

-> DeleteStackInstancesResponse

Create a value of DeleteStackInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:DeleteStackInstances', deleteStackInstancesResponse_operationId - The unique identifier for this stack set operation.

$sel:httpStatus:DeleteStackInstancesResponse', deleteStackInstancesResponse_httpStatus - The response's http status code.

Response Lenses

deleteStackInstancesResponse_operationId :: Lens' DeleteStackInstancesResponse (Maybe Text)

The unique identifier for this stack set operation.

deleteStackInstancesResponse_httpStatus :: Lens' DeleteStackInstancesResponse Int

The response's http status code.