Network.AWS.CloudFormation.DeregisterType

Description

Marks an extension or extension version as DEPRECATED in the CloudFormation registry, removing it from active use. Deprecated extensions or extension versions cannot be used in CloudFormation operations.

To deregister an entire extension, you must individually deregister all active versions of that extension. If an extension has only a single active version, deregistering that version results in the extension itself being deregistered and marked as deprecated in the registry.

You cannot deregister the default version of an extension if there are other active version of that extension. If you do deregister the default version of an extension, the textensionype itself is deregistered as well and marked as deprecated.

To view the deprecation status of an extension or extension version, use DescribeType.

Creating a Request

data DeregisterType

See: newDeregisterType smart constructor.

Constructors

DeregisterType'

Fields

versionId :: Maybe Text

The ID of a specific version of the extension. The version ID is the value at the end of the Amazon Resource Name (ARN) assigned to the extension version when it is registered.

typeName :: Maybe Text

The name of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

type' :: Maybe RegistryType

The kind of extension.

Conditional: You must specify either TypeName and Type, or Arn.

newDeregisterType :: DeregisterType

Create a value of DeregisterType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:DeregisterType', deregisterType_versionId - The ID of a specific version of the extension. The version ID is the value at the end of the Amazon Resource Name (ARN) assigned to the extension version when it is registered.

$sel:typeName:DeregisterType', deregisterType_typeName - The name of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

$sel:arn:DeregisterType', deregisterType_arn - The Amazon Resource Name (ARN) of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

$sel:type':DeregisterType', deregisterType_type - The kind of extension.

Conditional: You must specify either TypeName and Type, or Arn.

Request Lenses

deregisterType_versionId :: Lens' DeregisterType (Maybe Text)

The ID of a specific version of the extension. The version ID is the value at the end of the Amazon Resource Name (ARN) assigned to the extension version when it is registered.

deregisterType_typeName :: Lens' DeregisterType (Maybe Text)

The name of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

deregisterType_arn :: Lens' DeregisterType (Maybe Text)

The Amazon Resource Name (ARN) of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

deregisterType_type :: Lens' DeregisterType (Maybe RegistryType)

The kind of extension.

Conditional: You must specify either TypeName and Type, or Arn.

Destructuring the Response

data DeregisterTypeResponse

See: newDeregisterTypeResponse smart constructor.

Constructors

DeregisterTypeResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeregisterTypeResponse

Arguments

:: Int
$sel:httpStatus:DeregisterTypeResponse'

-> DeregisterTypeResponse

Create a value of DeregisterTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeregisterTypeResponse', deregisterTypeResponse_httpStatus - The response's http status code.

Response Lenses

deregisterTypeResponse_httpStatus :: Lens' DeregisterTypeResponse Int

The response's http status code.