Network.AWS.CloudFormation.DescribeChangeSet

Description

Returns the inputs for the change set and a list of changes that CloudFormation will make if you execute the change set. For more information, see Updating Stacks Using Change Sets in the CloudFormation User Guide.

This operation returns paginated results.

Creating a Request

data DescribeChangeSet

The input for the DescribeChangeSet action.

See: newDescribeChangeSet smart constructor.

Constructors

DescribeChangeSet'

Fields

nextToken :: Maybe Text

A string (provided by the DescribeChangeSet response output) that identifies the next page of information that you want to retrieve.

stackName :: Maybe Text

If you specified the name of a change set, specify the stack name or ID (ARN) of the change set you want to describe.

changeSetName :: Text

The name or Amazon Resource Name (ARN) of the change set that you want to describe.

newDescribeChangeSet

Arguments

:: Text
$sel:changeSetName:DescribeChangeSet'

-> DescribeChangeSet

Create a value of DescribeChangeSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeChangeSet', describeChangeSet_nextToken - A string (provided by the DescribeChangeSet response output) that identifies the next page of information that you want to retrieve.

$sel:stackName:DescribeChangeSet', describeChangeSet_stackName - If you specified the name of a change set, specify the stack name or ID (ARN) of the change set you want to describe.

$sel:changeSetName:DescribeChangeSet', describeChangeSet_changeSetName - The name or Amazon Resource Name (ARN) of the change set that you want to describe.

Request Lenses

describeChangeSet_nextToken :: Lens' DescribeChangeSet (Maybe Text)

A string (provided by the DescribeChangeSet response output) that identifies the next page of information that you want to retrieve.

describeChangeSet_stackName :: Lens' DescribeChangeSet (Maybe Text)

If you specified the name of a change set, specify the stack name or ID (ARN) of the change set you want to describe.

describeChangeSet_changeSetName :: Lens' DescribeChangeSet Text

The name or Amazon Resource Name (ARN) of the change set that you want to describe.

Destructuring the Response

data DescribeChangeSetResponse

The output for the DescribeChangeSet action.

See: newDescribeChangeSetResponse smart constructor.

Constructors

DescribeChangeSetResponse'

Fields

creationTime :: Maybe ISO8601

The start time when the change set was created, in UTC.

parentChangeSetId :: Maybe Text

Specifies the change set ID of the parent change set in the current nested change set hierarchy.

changes :: Maybe [Change]

A list of Change structures that describes the resources CloudFormation changes if you execute the change set.

notificationARNs :: Maybe [Text]

The ARNs of the Amazon Simple Notification Service (Amazon SNS) topics that will be associated with the stack if you execute the change set.

changeSetName :: Maybe Text

The name of the change set.

executionStatus :: Maybe ExecutionStatus

If the change set execution status is AVAILABLE, you can execute the change set. If you can’t execute the change set, the status indicates why. For example, a change set might be in an UNAVAILABLE state because CloudFormation is still creating it or in an OBSOLETE state because the stack was already updated.

changeSetId :: Maybe Text

The ARN of the change set.

includeNestedStacks :: Maybe Bool

Verifies if IncludeNestedStacks is set to True.

nextToken :: Maybe Text

If the output exceeds 1 MB, a string that identifies the next page of changes. If there is no additional page, this value is null.

rootChangeSetId :: Maybe Text

Specifies the change set ID of the root change set in the current nested change set hierarchy.

parameters :: Maybe [Parameter]

A list of Parameter structures that describes the input parameters and their values used to create the change set. For more information, see the Parameter data type.

statusReason :: Maybe Text

A description of the change set's status. For example, if your attempt to create a change set failed, CloudFormation shows the error message.

stackId :: Maybe Text

The ARN of the stack that is associated with the change set.

description :: Maybe Text

Information about the change set.

capabilities :: Maybe [Capability]

If you execute the change set, the list of capabilities that were explicitly acknowledged when the change set was created.

rollbackConfiguration :: Maybe RollbackConfiguration

The rollback triggers for CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.

tags :: Maybe [Tag]

If you execute the change set, the tags that will be associated with the stack.

stackName :: Maybe Text

The name of the stack that is associated with the change set.

httpStatus :: Int

The response's http status code.

status :: ChangeSetStatus

The current status of the change set, such as CREATE_IN_PROGRESS, CREATE_COMPLETE, or FAILED.

newDescribeChangeSetResponse

Arguments

:: Int
$sel:httpStatus:DescribeChangeSetResponse'

-> ChangeSetStatus
$sel:status:DescribeChangeSetResponse'

-> DescribeChangeSetResponse

Create a value of DescribeChangeSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeChangeSetResponse', describeChangeSetResponse_creationTime - The start time when the change set was created, in UTC.

$sel:parentChangeSetId:DescribeChangeSetResponse', describeChangeSetResponse_parentChangeSetId - Specifies the change set ID of the parent change set in the current nested change set hierarchy.

$sel:changes:DescribeChangeSetResponse', describeChangeSetResponse_changes - A list of Change structures that describes the resources CloudFormation changes if you execute the change set.

$sel:notificationARNs:DescribeChangeSetResponse', describeChangeSetResponse_notificationARNs - The ARNs of the Amazon Simple Notification Service (Amazon SNS) topics that will be associated with the stack if you execute the change set.

$sel:changeSetName:DescribeChangeSet', describeChangeSetResponse_changeSetName - The name of the change set.

$sel:executionStatus:DescribeChangeSetResponse', describeChangeSetResponse_executionStatus - If the change set execution status is AVAILABLE, you can execute the change set. If you can’t execute the change set, the status indicates why. For example, a change set might be in an UNAVAILABLE state because CloudFormation is still creating it or in an OBSOLETE state because the stack was already updated.

$sel:changeSetId:DescribeChangeSetResponse', describeChangeSetResponse_changeSetId - The ARN of the change set.

$sel:includeNestedStacks:DescribeChangeSetResponse', describeChangeSetResponse_includeNestedStacks - Verifies if IncludeNestedStacks is set to True.

$sel:nextToken:DescribeChangeSet', describeChangeSetResponse_nextToken - If the output exceeds 1 MB, a string that identifies the next page of changes. If there is no additional page, this value is null.

$sel:rootChangeSetId:DescribeChangeSetResponse', describeChangeSetResponse_rootChangeSetId - Specifies the change set ID of the root change set in the current nested change set hierarchy.

$sel:parameters:DescribeChangeSetResponse', describeChangeSetResponse_parameters - A list of Parameter structures that describes the input parameters and their values used to create the change set. For more information, see the Parameter data type.

$sel:statusReason:DescribeChangeSetResponse', describeChangeSetResponse_statusReason - A description of the change set's status. For example, if your attempt to create a change set failed, CloudFormation shows the error message.

$sel:stackId:DescribeChangeSetResponse', describeChangeSetResponse_stackId - The ARN of the stack that is associated with the change set.

$sel:description:DescribeChangeSetResponse', describeChangeSetResponse_description - Information about the change set.

$sel:capabilities:DescribeChangeSetResponse', describeChangeSetResponse_capabilities - If you execute the change set, the list of capabilities that were explicitly acknowledged when the change set was created.

$sel:rollbackConfiguration:DescribeChangeSetResponse', describeChangeSetResponse_rollbackConfiguration - The rollback triggers for CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.

$sel:tags:DescribeChangeSetResponse', describeChangeSetResponse_tags - If you execute the change set, the tags that will be associated with the stack.

$sel:stackName:DescribeChangeSet', describeChangeSetResponse_stackName - The name of the stack that is associated with the change set.

$sel:httpStatus:DescribeChangeSetResponse', describeChangeSetResponse_httpStatus - The response's http status code.

$sel:status:DescribeChangeSetResponse', describeChangeSetResponse_status - The current status of the change set, such as CREATE_IN_PROGRESS, CREATE_COMPLETE, or FAILED.

Response Lenses

describeChangeSetResponse_creationTime :: Lens' DescribeChangeSetResponse (Maybe UTCTime)

The start time when the change set was created, in UTC.

describeChangeSetResponse_parentChangeSetId :: Lens' DescribeChangeSetResponse (Maybe Text)

Specifies the change set ID of the parent change set in the current nested change set hierarchy.

describeChangeSetResponse_changes :: Lens' DescribeChangeSetResponse (Maybe [Change])

A list of Change structures that describes the resources CloudFormation changes if you execute the change set.

describeChangeSetResponse_notificationARNs :: Lens' DescribeChangeSetResponse (Maybe [Text])

The ARNs of the Amazon Simple Notification Service (Amazon SNS) topics that will be associated with the stack if you execute the change set.

describeChangeSetResponse_changeSetName :: Lens' DescribeChangeSetResponse (Maybe Text)

The name of the change set.

describeChangeSetResponse_executionStatus :: Lens' DescribeChangeSetResponse (Maybe ExecutionStatus)

If the change set execution status is AVAILABLE, you can execute the change set. If you can’t execute the change set, the status indicates why. For example, a change set might be in an UNAVAILABLE state because CloudFormation is still creating it or in an OBSOLETE state because the stack was already updated.

describeChangeSetResponse_changeSetId :: Lens' DescribeChangeSetResponse (Maybe Text)

The ARN of the change set.

describeChangeSetResponse_includeNestedStacks :: Lens' DescribeChangeSetResponse (Maybe Bool)

Verifies if IncludeNestedStacks is set to True.

describeChangeSetResponse_nextToken :: Lens' DescribeChangeSetResponse (Maybe Text)

If the output exceeds 1 MB, a string that identifies the next page of changes. If there is no additional page, this value is null.

describeChangeSetResponse_rootChangeSetId :: Lens' DescribeChangeSetResponse (Maybe Text)

Specifies the change set ID of the root change set in the current nested change set hierarchy.

describeChangeSetResponse_parameters :: Lens' DescribeChangeSetResponse (Maybe [Parameter])

A list of Parameter structures that describes the input parameters and their values used to create the change set. For more information, see the Parameter data type.

describeChangeSetResponse_statusReason :: Lens' DescribeChangeSetResponse (Maybe Text)

A description of the change set's status. For example, if your attempt to create a change set failed, CloudFormation shows the error message.

describeChangeSetResponse_stackId :: Lens' DescribeChangeSetResponse (Maybe Text)

The ARN of the stack that is associated with the change set.

describeChangeSetResponse_description :: Lens' DescribeChangeSetResponse (Maybe Text)

Information about the change set.

describeChangeSetResponse_capabilities :: Lens' DescribeChangeSetResponse (Maybe [Capability])

If you execute the change set, the list of capabilities that were explicitly acknowledged when the change set was created.

describeChangeSetResponse_rollbackConfiguration :: Lens' DescribeChangeSetResponse (Maybe RollbackConfiguration)

The rollback triggers for CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.

describeChangeSetResponse_tags :: Lens' DescribeChangeSetResponse (Maybe [Tag])

If you execute the change set, the tags that will be associated with the stack.

describeChangeSetResponse_stackName :: Lens' DescribeChangeSetResponse (Maybe Text)

The name of the stack that is associated with the change set.

describeChangeSetResponse_httpStatus :: Lens' DescribeChangeSetResponse Int

The response's http status code.

describeChangeSetResponse_status :: Lens' DescribeChangeSetResponse ChangeSetStatus

The current status of the change set, such as CREATE_IN_PROGRESS, CREATE_COMPLETE, or FAILED.