Network.AWS.CloudFormation.DescribeStackInstance

Description

Returns the stack instance that's associated with the specified stack set, Amazon Web Services account, and Region.

For a list of stack instances that are associated with a specific stack set, use ListStackInstances.

Creating a Request

data DescribeStackInstance

See: newDescribeStackInstance smart constructor.

Constructors

DescribeStackInstance'

Fields

callAs :: Maybe CallAs

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

stackSetName :: Text

The name or the unique stack ID of the stack set that you want to get stack instance information for.

stackInstanceAccount :: Text

The ID of an Amazon Web Services account that's associated with this stack instance.

stackInstanceRegion :: Text

The name of a Region that's associated with this stack instance.

newDescribeStackInstance

Arguments

:: Text
$sel:stackSetName:DescribeStackInstance'

-> Text
$sel:stackInstanceAccount:DescribeStackInstance'

-> Text
$sel:stackInstanceRegion:DescribeStackInstance'

-> DescribeStackInstance

Create a value of DescribeStackInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:callAs:DescribeStackInstance', describeStackInstance_callAs - [Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

$sel:stackSetName:DescribeStackInstance', describeStackInstance_stackSetName - The name or the unique stack ID of the stack set that you want to get stack instance information for.

$sel:stackInstanceAccount:DescribeStackInstance', describeStackInstance_stackInstanceAccount - The ID of an Amazon Web Services account that's associated with this stack instance.

$sel:stackInstanceRegion:DescribeStackInstance', describeStackInstance_stackInstanceRegion - The name of a Region that's associated with this stack instance.

Request Lenses

describeStackInstance_callAs :: Lens' DescribeStackInstance (Maybe CallAs)

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

describeStackInstance_stackSetName :: Lens' DescribeStackInstance Text

The name or the unique stack ID of the stack set that you want to get stack instance information for.

describeStackInstance_stackInstanceAccount :: Lens' DescribeStackInstance Text

The ID of an Amazon Web Services account that's associated with this stack instance.

describeStackInstance_stackInstanceRegion :: Lens' DescribeStackInstance Text

The name of a Region that's associated with this stack instance.

Destructuring the Response

data DescribeStackInstanceResponse

See: newDescribeStackInstanceResponse smart constructor.

Constructors

DescribeStackInstanceResponse'

Fields

stackInstance :: Maybe StackInstance

The stack instance that matches the specified request parameters.

httpStatus :: Int

The response's http status code.

newDescribeStackInstanceResponse

Arguments

:: Int
$sel:httpStatus:DescribeStackInstanceResponse'

-> DescribeStackInstanceResponse

Create a value of DescribeStackInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackInstance:DescribeStackInstanceResponse', describeStackInstanceResponse_stackInstance - The stack instance that matches the specified request parameters.

$sel:httpStatus:DescribeStackInstanceResponse', describeStackInstanceResponse_httpStatus - The response's http status code.

Response Lenses

describeStackInstanceResponse_stackInstance :: Lens' DescribeStackInstanceResponse (Maybe StackInstance)

The stack instance that matches the specified request parameters.

describeStackInstanceResponse_httpStatus :: Lens' DescribeStackInstanceResponse Int

The response's http status code.