Network.AWS.CloudFormation.DescribeStackResource

Description

Returns a description of the specified resource in the specified stack.

For deleted stacks, DescribeStackResource returns resource information for up to 90 days after the stack has been deleted.

Creating a Request

data DescribeStackResource

The input for DescribeStackResource action.

See: newDescribeStackResource smart constructor.

Constructors

DescribeStackResource'

Fields

stackName :: Text

The name or the unique stack ID that is associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID.

Deleted stacks: You must specify the unique stack ID.

Default: There is no default value.

logicalResourceId :: Text

The logical name of the resource as specified in the template.

Default: There is no default value.

newDescribeStackResource

Arguments

:: Text
$sel:stackName:DescribeStackResource'

-> Text
$sel:logicalResourceId:DescribeStackResource'

-> DescribeStackResource

Create a value of DescribeStackResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackName:DescribeStackResource', describeStackResource_stackName - The name or the unique stack ID that is associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID.

Deleted stacks: You must specify the unique stack ID.

Default: There is no default value.

$sel:logicalResourceId:DescribeStackResource', describeStackResource_logicalResourceId - The logical name of the resource as specified in the template.

Default: There is no default value.

Request Lenses

describeStackResource_stackName :: Lens' DescribeStackResource Text

The name or the unique stack ID that is associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID.

Deleted stacks: You must specify the unique stack ID.

Default: There is no default value.

describeStackResource_logicalResourceId :: Lens' DescribeStackResource Text

The logical name of the resource as specified in the template.

Default: There is no default value.

Destructuring the Response

data DescribeStackResourceResponse

The output for a DescribeStackResource action.

See: newDescribeStackResourceResponse smart constructor.

Constructors

DescribeStackResourceResponse'

Fields

stackResourceDetail :: Maybe StackResourceDetail

A StackResourceDetail structure containing the description of the specified resource in the specified stack.

httpStatus :: Int

The response's http status code.

newDescribeStackResourceResponse

Arguments

:: Int
$sel:httpStatus:DescribeStackResourceResponse'

-> DescribeStackResourceResponse

Create a value of DescribeStackResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackResourceDetail:DescribeStackResourceResponse', describeStackResourceResponse_stackResourceDetail - A StackResourceDetail structure containing the description of the specified resource in the specified stack.

$sel:httpStatus:DescribeStackResourceResponse', describeStackResourceResponse_httpStatus - The response's http status code.

Response Lenses

describeStackResourceResponse_stackResourceDetail :: Lens' DescribeStackResourceResponse (Maybe StackResourceDetail)

A StackResourceDetail structure containing the description of the specified resource in the specified stack.

describeStackResourceResponse_httpStatus :: Lens' DescribeStackResourceResponse Int

The response's http status code.