Network.AWS.CloudFormation.DescribeStackResourceDrifts

Description

Returns drift information for the resources that have been checked for drift in the specified stack. This includes actual and expected configuration values for resources where CloudFormation detects configuration drift.

For a given stack, there will be one StackResourceDrift for each stack resource that has been checked for drift. Resources that haven't yet been checked for drift are not included. Resources that do not currently support drift detection are not checked, and so not included. For a list of resources that support drift detection, see Resources that Support Drift Detection.

Use DetectStackResourceDrift to detect drift on individual resources, or DetectStackDrift to detect drift on all supported resources for a given stack.

Creating a Request

data DescribeStackResourceDrifts

See: newDescribeStackResourceDrifts smart constructor.

Constructors

DescribeStackResourceDrifts'

Fields

nextToken :: Maybe Text

A string that identifies the next page of stack resource drift results.

maxResults :: Maybe Natural

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

stackResourceDriftStatusFilters :: Maybe (NonEmpty StackResourceDriftStatus)

The resource drift status values to use as filters for the resource drift results returned.

DELETED: The resource differs from its expected template configuration in that the resource has been deleted.

MODIFIED: One or more resource properties differ from their expected template values.

IN_SYNC: The resources's actual configuration matches its expected template configuration.

NOT_CHECKED: CloudFormation does not currently return this value.

stackName :: Text

The name of the stack for which you want drift information.

newDescribeStackResourceDrifts

Arguments

:: Text
$sel:stackName:DescribeStackResourceDrifts'

-> DescribeStackResourceDrifts

Create a value of DescribeStackResourceDrifts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeStackResourceDrifts', describeStackResourceDrifts_nextToken - A string that identifies the next page of stack resource drift results.

$sel:maxResults:DescribeStackResourceDrifts', describeStackResourceDrifts_maxResults - The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

$sel:stackResourceDriftStatusFilters:DescribeStackResourceDrifts', describeStackResourceDrifts_stackResourceDriftStatusFilters - The resource drift status values to use as filters for the resource drift results returned.

DELETED: The resource differs from its expected template configuration in that the resource has been deleted.

MODIFIED: One or more resource properties differ from their expected template values.

IN_SYNC: The resources's actual configuration matches its expected template configuration.

NOT_CHECKED: CloudFormation does not currently return this value.

$sel:stackName:DescribeStackResourceDrifts', describeStackResourceDrifts_stackName - The name of the stack for which you want drift information.

Request Lenses

describeStackResourceDrifts_nextToken :: Lens' DescribeStackResourceDrifts (Maybe Text)

A string that identifies the next page of stack resource drift results.

describeStackResourceDrifts_maxResults :: Lens' DescribeStackResourceDrifts (Maybe Natural)

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

describeStackResourceDrifts_stackResourceDriftStatusFilters :: Lens' DescribeStackResourceDrifts (Maybe (NonEmpty StackResourceDriftStatus))

The resource drift status values to use as filters for the resource drift results returned.

DELETED: The resource differs from its expected template configuration in that the resource has been deleted.

MODIFIED: One or more resource properties differ from their expected template values.

IN_SYNC: The resources's actual configuration matches its expected template configuration.

NOT_CHECKED: CloudFormation does not currently return this value.

describeStackResourceDrifts_stackName :: Lens' DescribeStackResourceDrifts Text

The name of the stack for which you want drift information.

Destructuring the Response

data DescribeStackResourceDriftsResponse

See: newDescribeStackResourceDriftsResponse smart constructor.

Constructors

DescribeStackResourceDriftsResponse'

Fields

nextToken :: Maybe Text

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call DescribeStackResourceDrifts again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

httpStatus :: Int

The response's http status code.

stackResourceDrifts :: [StackResourceDrift]

Drift information for the resources that have been checked for drift in the specified stack. This includes actual and expected configuration values for resources where CloudFormation detects drift.

For a given stack, there will be one StackResourceDrift for each stack resource that has been checked for drift. Resources that have not yet been checked for drift are not included. Resources that do not currently support drift detection are not checked, and so not included. For a list of resources that support drift detection, see Resources that Support Drift Detection.

newDescribeStackResourceDriftsResponse

Arguments

:: Int
$sel:httpStatus:DescribeStackResourceDriftsResponse'

-> DescribeStackResourceDriftsResponse

Create a value of DescribeStackResourceDriftsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeStackResourceDrifts', describeStackResourceDriftsResponse_nextToken - If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call DescribeStackResourceDrifts again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

$sel:httpStatus:DescribeStackResourceDriftsResponse', describeStackResourceDriftsResponse_httpStatus - The response's http status code.

$sel:stackResourceDrifts:DescribeStackResourceDriftsResponse', describeStackResourceDriftsResponse_stackResourceDrifts - Drift information for the resources that have been checked for drift in the specified stack. This includes actual and expected configuration values for resources where CloudFormation detects drift.

For a given stack, there will be one StackResourceDrift for each stack resource that has been checked for drift. Resources that have not yet been checked for drift are not included. Resources that do not currently support drift detection are not checked, and so not included. For a list of resources that support drift detection, see Resources that Support Drift Detection.

Response Lenses

describeStackResourceDriftsResponse_nextToken :: Lens' DescribeStackResourceDriftsResponse (Maybe Text)

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call DescribeStackResourceDrifts again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

describeStackResourceDriftsResponse_httpStatus :: Lens' DescribeStackResourceDriftsResponse Int

The response's http status code.

describeStackResourceDriftsResponse_stackResourceDrifts :: Lens' DescribeStackResourceDriftsResponse [StackResourceDrift]

Drift information for the resources that have been checked for drift in the specified stack. This includes actual and expected configuration values for resources where CloudFormation detects drift.

For a given stack, there will be one StackResourceDrift for each stack resource that has been checked for drift. Resources that have not yet been checked for drift are not included. Resources that do not currently support drift detection are not checked, and so not included. For a list of resources that support drift detection, see Resources that Support Drift Detection.