Network.AWS.CloudFormation.DescribeStackResources

Description

Returns Amazon Web Services resource descriptions for running and deleted stacks. If StackName is specified, all the associated resources that are part of the stack are returned. If PhysicalResourceId is specified, the associated resources of the stack that the resource belongs to are returned.

Only the first 100 resources will be returned. If your stack has more resources than this, you should use ListStackResources instead.

For deleted stacks, DescribeStackResources returns resource information for up to 90 days after the stack has been deleted.

You must specify either StackName or PhysicalResourceId, but not both. In addition, you can specify LogicalResourceId to filter the returned result. For more information about resources, the LogicalResourceId and PhysicalResourceId, go to the CloudFormation User Guide.

A ValidationError is returned if you specify both StackName and PhysicalResourceId in the same request.

Creating a Request

data DescribeStackResources

The input for DescribeStackResources action.

See: newDescribeStackResources smart constructor.

Constructors

DescribeStackResources'

Fields

stackName :: Maybe Text

The name or the unique stack ID that is associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID.

Deleted stacks: You must specify the unique stack ID.

Default: There is no default value.

Required: Conditional. If you do not specify StackName, you must specify PhysicalResourceId.

physicalResourceId :: Maybe Text

The name or unique identifier that corresponds to a physical instance ID of a resource supported by CloudFormation.

For example, for an Amazon Elastic Compute Cloud (EC2) instance, PhysicalResourceId corresponds to the InstanceId. You can pass the EC2 InstanceId to DescribeStackResources to find which stack the instance belongs to and what other resources are part of the stack.

Required: Conditional. If you do not specify PhysicalResourceId, you must specify StackName.

Default: There is no default value.

logicalResourceId :: Maybe Text

The logical name of the resource as specified in the template.

Default: There is no default value.

newDescribeStackResources :: DescribeStackResources

Create a value of DescribeStackResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackName:DescribeStackResources', describeStackResources_stackName - The name or the unique stack ID that is associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID.

Deleted stacks: You must specify the unique stack ID.

Default: There is no default value.

Required: Conditional. If you do not specify StackName, you must specify PhysicalResourceId.

$sel:physicalResourceId:DescribeStackResources', describeStackResources_physicalResourceId - The name or unique identifier that corresponds to a physical instance ID of a resource supported by CloudFormation.

For example, for an Amazon Elastic Compute Cloud (EC2) instance, PhysicalResourceId corresponds to the InstanceId. You can pass the EC2 InstanceId to DescribeStackResources to find which stack the instance belongs to and what other resources are part of the stack.

Required: Conditional. If you do not specify PhysicalResourceId, you must specify StackName.

Default: There is no default value.

$sel:logicalResourceId:DescribeStackResources', describeStackResources_logicalResourceId - The logical name of the resource as specified in the template.

Default: There is no default value.

Request Lenses

describeStackResources_stackName :: Lens' DescribeStackResources (Maybe Text)

The name or the unique stack ID that is associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID.

Deleted stacks: You must specify the unique stack ID.

Default: There is no default value.

Required: Conditional. If you do not specify StackName, you must specify PhysicalResourceId.

describeStackResources_physicalResourceId :: Lens' DescribeStackResources (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of a resource supported by CloudFormation.

For example, for an Amazon Elastic Compute Cloud (EC2) instance, PhysicalResourceId corresponds to the InstanceId. You can pass the EC2 InstanceId to DescribeStackResources to find which stack the instance belongs to and what other resources are part of the stack.

Required: Conditional. If you do not specify PhysicalResourceId, you must specify StackName.

Default: There is no default value.

describeStackResources_logicalResourceId :: Lens' DescribeStackResources (Maybe Text)

The logical name of the resource as specified in the template.

Default: There is no default value.

Destructuring the Response

data DescribeStackResourcesResponse

The output for a DescribeStackResources action.

See: newDescribeStackResourcesResponse smart constructor.

Constructors

DescribeStackResourcesResponse'

Fields

stackResources :: Maybe [StackResource]

A list of StackResource structures.

httpStatus :: Int

The response's http status code.

newDescribeStackResourcesResponse

Arguments

:: Int
$sel:httpStatus:DescribeStackResourcesResponse'

-> DescribeStackResourcesResponse

Create a value of DescribeStackResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackResources:DescribeStackResourcesResponse', describeStackResourcesResponse_stackResources - A list of StackResource structures.

$sel:httpStatus:DescribeStackResourcesResponse', describeStackResourcesResponse_httpStatus - The response's http status code.

Response Lenses

describeStackResourcesResponse_stackResources :: Lens' DescribeStackResourcesResponse (Maybe [StackResource])

A list of StackResource structures.

describeStackResourcesResponse_httpStatus :: Lens' DescribeStackResourcesResponse Int

The response's http status code.