Network.AWS.CloudFormation.DescribeStackSetOperation

Description

Returns the description of the specified stack set operation.

Creating a Request

data DescribeStackSetOperation

See: newDescribeStackSetOperation smart constructor.

Constructors

DescribeStackSetOperation'

Fields

callAs :: Maybe CallAs

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

stackSetName :: Text

The name or the unique stack ID of the stack set for the stack operation.

operationId :: Text

The unique ID of the stack set operation.

newDescribeStackSetOperation

Arguments

:: Text
$sel:stackSetName:DescribeStackSetOperation'

-> Text
$sel:operationId:DescribeStackSetOperation'

-> DescribeStackSetOperation

Create a value of DescribeStackSetOperation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:callAs:DescribeStackSetOperation', describeStackSetOperation_callAs - [Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

$sel:stackSetName:DescribeStackSetOperation', describeStackSetOperation_stackSetName - The name or the unique stack ID of the stack set for the stack operation.

$sel:operationId:DescribeStackSetOperation', describeStackSetOperation_operationId - The unique ID of the stack set operation.

Request Lenses

describeStackSetOperation_callAs :: Lens' DescribeStackSetOperation (Maybe CallAs)

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

describeStackSetOperation_stackSetName :: Lens' DescribeStackSetOperation Text

The name or the unique stack ID of the stack set for the stack operation.

describeStackSetOperation_operationId :: Lens' DescribeStackSetOperation Text

The unique ID of the stack set operation.

Destructuring the Response

data DescribeStackSetOperationResponse

See: newDescribeStackSetOperationResponse smart constructor.

Constructors

DescribeStackSetOperationResponse'

Fields

stackSetOperation :: Maybe StackSetOperation

The specified stack set operation.

httpStatus :: Int

The response's http status code.

newDescribeStackSetOperationResponse

Arguments

:: Int
$sel:httpStatus:DescribeStackSetOperationResponse'

-> DescribeStackSetOperationResponse

Create a value of DescribeStackSetOperationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackSetOperation:DescribeStackSetOperationResponse', describeStackSetOperationResponse_stackSetOperation - The specified stack set operation.

$sel:httpStatus:DescribeStackSetOperationResponse', describeStackSetOperationResponse_httpStatus - The response's http status code.

Response Lenses

describeStackSetOperationResponse_stackSetOperation :: Lens' DescribeStackSetOperationResponse (Maybe StackSetOperation)

The specified stack set operation.

describeStackSetOperationResponse_httpStatus :: Lens' DescribeStackSetOperationResponse Int

The response's http status code.