Network.AWS.CloudFormation.DescribeType

Description

Returns detailed information about an extension that has been registered.

If you specify a VersionId, DescribeType returns information about that specific extension version. Otherwise, it returns information about the default extension version.

Creating a Request

data DescribeType

See: newDescribeType smart constructor.

Constructors

DescribeType'

Fields

typeName :: Maybe Text

The name of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

publisherId :: Maybe Text

The publisher ID of the extension publisher.

Extensions provided by Amazon are not assigned a publisher ID.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

versionId :: Maybe Text

The ID of a specific version of the extension. The version ID is the value at the end of the Amazon Resource Name (ARN) assigned to the extension version when it is registered.

If you specify a VersionId, DescribeType returns information about that specific extension version. Otherwise, it returns information about the default extension version.

publicVersionNumber :: Maybe Text

The version number of a public third-party extension.

type' :: Maybe RegistryType

The kind of extension.

Conditional: You must specify either TypeName and Type, or Arn.

newDescribeType :: DescribeType

Create a value of DescribeType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:DescribeType', describeType_typeName - The name of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

$sel:publisherId:DescribeType', describeType_publisherId - The publisher ID of the extension publisher.

Extensions provided by Amazon are not assigned a publisher ID.

$sel:arn:DescribeType', describeType_arn - The Amazon Resource Name (ARN) of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

$sel:versionId:DescribeType', describeType_versionId - The ID of a specific version of the extension. The version ID is the value at the end of the Amazon Resource Name (ARN) assigned to the extension version when it is registered.

If you specify a VersionId, DescribeType returns information about that specific extension version. Otherwise, it returns information about the default extension version.

$sel:publicVersionNumber:DescribeType', describeType_publicVersionNumber - The version number of a public third-party extension.

$sel:type':DescribeType', describeType_type - The kind of extension.

Conditional: You must specify either TypeName and Type, or Arn.

Request Lenses

describeType_typeName :: Lens' DescribeType (Maybe Text)

The name of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

describeType_publisherId :: Lens' DescribeType (Maybe Text)

The publisher ID of the extension publisher.

Extensions provided by Amazon are not assigned a publisher ID.

describeType_arn :: Lens' DescribeType (Maybe Text)

The Amazon Resource Name (ARN) of the extension.

Conditional: You must specify either TypeName and Type, or Arn.

describeType_versionId :: Lens' DescribeType (Maybe Text)

The ID of a specific version of the extension. The version ID is the value at the end of the Amazon Resource Name (ARN) assigned to the extension version when it is registered.

If you specify a VersionId, DescribeType returns information about that specific extension version. Otherwise, it returns information about the default extension version.

describeType_publicVersionNumber :: Lens' DescribeType (Maybe Text)

The version number of a public third-party extension.

describeType_type :: Lens' DescribeType (Maybe RegistryType)

The kind of extension.

Conditional: You must specify either TypeName and Type, or Arn.

Destructuring the Response

data DescribeTypeResponse

See: newDescribeTypeResponse smart constructor.

Constructors

DescribeTypeResponse'

Fields

isActivated :: Maybe Bool

Whether or not the extension is activated in the account and region.

This only applies to public third-party extensions. For all other extensions, CloudFormation returns null.

typeTestsStatusDescription :: Maybe Text

The description of the test status. To return the extension test status of a specifc extension version, you must specify VersionId.

This applies only to registered private extension versions. CloudFormation does not return this information for public extensions, whether or not they are activated in your account.

typeName :: Maybe Text

The name of the extension.

If the extension is a public third-party type you have activated with a type name alias, CloudFormation returns the type name alias. For more information, see ActivateType.

originalTypeArn :: Maybe Text

For public extensions that have been activated for this account and region, the Amazon Resource Name (ARN) of the public extension.

loggingConfig :: Maybe LoggingConfig

Contains logging configuration information for private extensions. This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

typeTestsStatus :: Maybe TypeTestsStatus

The contract test status of the registered extension version. To return the extension test status of a specifc extension version, you must specify VersionId.

This applies only to registered private extension versions. CloudFormation does not return this information for public extensions, whether or not they are activated in your account.

PASSED: The extension has passed all its contract tests.

An extension must have a test status of PASSED before it can be published. For more information, see Publishing extensions to make them available for public use in the CloudFormation Command Line Interface User Guide.

FAILED: The extension has failed one or more contract tests.

IN_PROGRESS: Contract tests are currently being performed on the extension.

NOT_TESTED: Contract tests have not been performed on the extension.

configurationSchema :: Maybe Text

A JSON string that represent the current configuration data for the extension in this account and region.

To set the configuration data for an extension, use SetTypeConfiguration. For more information, see Configuring extensions at the account level in the CloudFormation User Guide.

schema :: Maybe Text

The schema that defines the extension.

For more information on extension schemas, see Resource Provider Schema in the CloudFormation CLI User Guide.

publisherId :: Maybe Text

The publisher ID of the extension publisher.

This applies only to public third-party extensions. For private registered extensions, and extensions provided by Amazon, CloudFormation returns null.

executionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM execution role used to register the extension. This applies only to private extensions you have registered in your account. For more information, see RegisterType.

If the registered extension calls any Amazon Web Services APIs, you must create an <https:docs.aws.amazon.comIAMlatestUserGuideid_roles.html IAM execution role> that includes the necessary permissions to call those Amazon Web Services APIs, and provision that execution role in your account. CloudFormation then assumes that execution role to provide your extension with the appropriate credentials.

latestPublicVersion :: Maybe Text

The latest version of a public extension that is available for use.

This only applies if you specify a public extension, and you do not specify a version. For all other requests, CloudFormation returns null.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the extension.

deprecatedStatus :: Maybe DeprecatedStatus

The deprecation status of the extension version.

Valid values include:

LIVE: The extension is activated or registered and can be used in CloudFormation operations, dependent on its provisioning behavior and visibility scope.

DEPRECATED: The extension has been deactivated or deregistered and can no longer be used in CloudFormation operations.

For public third-party extensions, CloudFormation returns null.

requiredActivatedTypes :: Maybe [RequiredActivatedType]

For extensions that are modules, the public third-party extensions that must be activated in your account in order for the module itself to be activated.

lastUpdated :: Maybe ISO8601

When the specified extension version was registered. This applies only to:

Private extensions you have registered in your account. For more information, see RegisterType.

Public extensions you have activated in your account with auto-update specified. For more information, see ActivateType.

defaultVersionId :: Maybe Text

The ID of the default version of the extension. The default version is used when the extension version is not specified.

This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

To set the default version of an extension, use SetTypeDefaultVersion .

documentationUrl :: Maybe Text

The URL of a page providing detailed documentation for this extension.

provisioningType :: Maybe ProvisioningType

For resource type extensions, the provisioning behavior of the resource type. CloudFormation determines the provisioning type during registration, based on the types of handlers in the schema handler package submitted.

Valid values include:

FULLY_MUTABLE: The resource type includes an update handler to process updates to the type during stack update operations.

IMMUTABLE: The resource type does not include an update handler, so the type cannot be updated and must instead be replaced during stack update operations.

NON_PROVISIONABLE: The resource type does not include all of the following handlers, and therefore cannot actually be provisioned.

create

read

delete

sourceUrl :: Maybe Text

The URL of the source code for the extension.

description :: Maybe Text

The description of the extension.

autoUpdate :: Maybe Bool

Whether CloudFormation automatically updates the extension in this account and region when a new minor version is published by the extension publisher. Major versions released by the publisher must be manually updated. For more information, see Activating public extensions for use in your account in the CloudFormation User Guide.

originalTypeName :: Maybe Text

For public extensions that have been activated for this account and region, the type name of the public extension.

If you specified a TypeNameAlias when enabling the extension in this account and region, CloudFormation treats that alias as the extension's type name within the account and region, not the type name of the public extension. For more information, see Specifying aliases to refer to extensions in the CloudFormation User Guide.

visibility :: Maybe Visibility

The scope at which the extension is visible and usable in CloudFormation operations.

Valid values include:

PRIVATE: The extension is only visible and usable within the account in which it is registered. CloudFormation marks any extensions you register as PRIVATE.

PUBLIC: The extension is publically visible and usable within any Amazon account.

isDefaultVersion :: Maybe Bool

Whether the specified extension version is set as the default version.

This applies only to private extensions you have registered in your account, and extensions published by Amazon. For public third-party extensions, whether or not they are activated in your account, CloudFormation returns null.

publicVersionNumber :: Maybe Text

The version number of a public third-party extension.

This applies only if you specify a public extension you have activated in your account, or specify a public extension without specifying a version. For all other extensions, CloudFormation returns null.

type' :: Maybe RegistryType

The kind of extension.

timeCreated :: Maybe ISO8601

When the specified private extension version was registered or activated in your account.

httpStatus :: Int

The response's http status code.

newDescribeTypeResponse

Arguments

:: Int
$sel:httpStatus:DescribeTypeResponse'

-> DescribeTypeResponse

Create a value of DescribeTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isActivated:DescribeTypeResponse', describeTypeResponse_isActivated - Whether or not the extension is activated in the account and region.

This only applies to public third-party extensions. For all other extensions, CloudFormation returns null.

$sel:typeTestsStatusDescription:DescribeTypeResponse', describeTypeResponse_typeTestsStatusDescription - The description of the test status. To return the extension test status of a specifc extension version, you must specify VersionId.

This applies only to registered private extension versions. CloudFormation does not return this information for public extensions, whether or not they are activated in your account.

$sel:typeName:DescribeType', describeTypeResponse_typeName - The name of the extension.

If the extension is a public third-party type you have activated with a type name alias, CloudFormation returns the type name alias. For more information, see ActivateType.

$sel:originalTypeArn:DescribeTypeResponse', describeTypeResponse_originalTypeArn - For public extensions that have been activated for this account and region, the Amazon Resource Name (ARN) of the public extension.

$sel:loggingConfig:DescribeTypeResponse', describeTypeResponse_loggingConfig - Contains logging configuration information for private extensions. This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

$sel:typeTestsStatus:DescribeTypeResponse', describeTypeResponse_typeTestsStatus - The contract test status of the registered extension version. To return the extension test status of a specifc extension version, you must specify VersionId.

This applies only to registered private extension versions. CloudFormation does not return this information for public extensions, whether or not they are activated in your account.

PASSED: The extension has passed all its contract tests.

An extension must have a test status of PASSED before it can be published. For more information, see Publishing extensions to make them available for public use in the CloudFormation Command Line Interface User Guide.

FAILED: The extension has failed one or more contract tests.

IN_PROGRESS: Contract tests are currently being performed on the extension.

NOT_TESTED: Contract tests have not been performed on the extension.

$sel:configurationSchema:DescribeTypeResponse', describeTypeResponse_configurationSchema - A JSON string that represent the current configuration data for the extension in this account and region.

To set the configuration data for an extension, use SetTypeConfiguration. For more information, see Configuring extensions at the account level in the CloudFormation User Guide.

$sel:schema:DescribeTypeResponse', describeTypeResponse_schema - The schema that defines the extension.

For more information on extension schemas, see Resource Provider Schema in the CloudFormation CLI User Guide.

$sel:publisherId:DescribeType', describeTypeResponse_publisherId - The publisher ID of the extension publisher.

This applies only to public third-party extensions. For private registered extensions, and extensions provided by Amazon, CloudFormation returns null.

$sel:executionRoleArn:DescribeTypeResponse', describeTypeResponse_executionRoleArn - The Amazon Resource Name (ARN) of the IAM execution role used to register the extension. This applies only to private extensions you have registered in your account. For more information, see RegisterType.

If the registered extension calls any Amazon Web Services APIs, you must create an <https:docs.aws.amazon.comIAMlatestUserGuideid_roles.html IAM execution role> that includes the necessary permissions to call those Amazon Web Services APIs, and provision that execution role in your account. CloudFormation then assumes that execution role to provide your extension with the appropriate credentials.

$sel:latestPublicVersion:DescribeTypeResponse', describeTypeResponse_latestPublicVersion - The latest version of a public extension that is available for use.

This only applies if you specify a public extension, and you do not specify a version. For all other requests, CloudFormation returns null.

$sel:arn:DescribeType', describeTypeResponse_arn - The Amazon Resource Name (ARN) of the extension.

$sel:deprecatedStatus:DescribeTypeResponse', describeTypeResponse_deprecatedStatus - The deprecation status of the extension version.

Valid values include:

LIVE: The extension is activated or registered and can be used in CloudFormation operations, dependent on its provisioning behavior and visibility scope.

DEPRECATED: The extension has been deactivated or deregistered and can no longer be used in CloudFormation operations.

For public third-party extensions, CloudFormation returns null.

$sel:requiredActivatedTypes:DescribeTypeResponse', describeTypeResponse_requiredActivatedTypes - For extensions that are modules, the public third-party extensions that must be activated in your account in order for the module itself to be activated.

$sel:lastUpdated:DescribeTypeResponse', describeTypeResponse_lastUpdated - When the specified extension version was registered. This applies only to:

Private extensions you have registered in your account. For more information, see RegisterType.

Public extensions you have activated in your account with auto-update specified. For more information, see ActivateType.

$sel:defaultVersionId:DescribeTypeResponse', describeTypeResponse_defaultVersionId - The ID of the default version of the extension. The default version is used when the extension version is not specified.

This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

To set the default version of an extension, use SetTypeDefaultVersion .

$sel:documentationUrl:DescribeTypeResponse', describeTypeResponse_documentationUrl - The URL of a page providing detailed documentation for this extension.

$sel:provisioningType:DescribeTypeResponse', describeTypeResponse_provisioningType - For resource type extensions, the provisioning behavior of the resource type. CloudFormation determines the provisioning type during registration, based on the types of handlers in the schema handler package submitted.

Valid values include:

FULLY_MUTABLE: The resource type includes an update handler to process updates to the type during stack update operations.

IMMUTABLE: The resource type does not include an update handler, so the type cannot be updated and must instead be replaced during stack update operations.

NON_PROVISIONABLE: The resource type does not include all of the following handlers, and therefore cannot actually be provisioned.

create

read

delete

$sel:sourceUrl:DescribeTypeResponse', describeTypeResponse_sourceUrl - The URL of the source code for the extension.

$sel:description:DescribeTypeResponse', describeTypeResponse_description - The description of the extension.

$sel:autoUpdate:DescribeTypeResponse', describeTypeResponse_autoUpdate - Whether CloudFormation automatically updates the extension in this account and region when a new minor version is published by the extension publisher. Major versions released by the publisher must be manually updated. For more information, see Activating public extensions for use in your account in the CloudFormation User Guide.

$sel:originalTypeName:DescribeTypeResponse', describeTypeResponse_originalTypeName - For public extensions that have been activated for this account and region, the type name of the public extension.

If you specified a TypeNameAlias when enabling the extension in this account and region, CloudFormation treats that alias as the extension's type name within the account and region, not the type name of the public extension. For more information, see Specifying aliases to refer to extensions in the CloudFormation User Guide.

$sel:visibility:DescribeTypeResponse', describeTypeResponse_visibility - The scope at which the extension is visible and usable in CloudFormation operations.

Valid values include:

PRIVATE: The extension is only visible and usable within the account in which it is registered. CloudFormation marks any extensions you register as PRIVATE.

PUBLIC: The extension is publically visible and usable within any Amazon account.

$sel:isDefaultVersion:DescribeTypeResponse', describeTypeResponse_isDefaultVersion - Whether the specified extension version is set as the default version.

This applies only to private extensions you have registered in your account, and extensions published by Amazon. For public third-party extensions, whether or not they are activated in your account, CloudFormation returns null.

$sel:publicVersionNumber:DescribeType', describeTypeResponse_publicVersionNumber - The version number of a public third-party extension.

This applies only if you specify a public extension you have activated in your account, or specify a public extension without specifying a version. For all other extensions, CloudFormation returns null.

$sel:type':DescribeType', describeTypeResponse_type - The kind of extension.

$sel:timeCreated:DescribeTypeResponse', describeTypeResponse_timeCreated - When the specified private extension version was registered or activated in your account.

$sel:httpStatus:DescribeTypeResponse', describeTypeResponse_httpStatus - The response's http status code.

Response Lenses

describeTypeResponse_isActivated :: Lens' DescribeTypeResponse (Maybe Bool)

Whether or not the extension is activated in the account and region.

This only applies to public third-party extensions. For all other extensions, CloudFormation returns null.

describeTypeResponse_typeTestsStatusDescription :: Lens' DescribeTypeResponse (Maybe Text)

The description of the test status. To return the extension test status of a specifc extension version, you must specify VersionId.

This applies only to registered private extension versions. CloudFormation does not return this information for public extensions, whether or not they are activated in your account.

describeTypeResponse_typeName :: Lens' DescribeTypeResponse (Maybe Text)

The name of the extension.

If the extension is a public third-party type you have activated with a type name alias, CloudFormation returns the type name alias. For more information, see ActivateType.

describeTypeResponse_originalTypeArn :: Lens' DescribeTypeResponse (Maybe Text)

For public extensions that have been activated for this account and region, the Amazon Resource Name (ARN) of the public extension.

describeTypeResponse_loggingConfig :: Lens' DescribeTypeResponse (Maybe LoggingConfig)

Contains logging configuration information for private extensions. This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

describeTypeResponse_typeTestsStatus :: Lens' DescribeTypeResponse (Maybe TypeTestsStatus)

The contract test status of the registered extension version. To return the extension test status of a specifc extension version, you must specify VersionId.

This applies only to registered private extension versions. CloudFormation does not return this information for public extensions, whether or not they are activated in your account.

PASSED: The extension has passed all its contract tests.

An extension must have a test status of PASSED before it can be published. For more information, see Publishing extensions to make them available for public use in the CloudFormation Command Line Interface User Guide.

FAILED: The extension has failed one or more contract tests.

IN_PROGRESS: Contract tests are currently being performed on the extension.

NOT_TESTED: Contract tests have not been performed on the extension.

describeTypeResponse_configurationSchema :: Lens' DescribeTypeResponse (Maybe Text)

A JSON string that represent the current configuration data for the extension in this account and region.

To set the configuration data for an extension, use SetTypeConfiguration. For more information, see Configuring extensions at the account level in the CloudFormation User Guide.

describeTypeResponse_schema :: Lens' DescribeTypeResponse (Maybe Text)

The schema that defines the extension.

For more information on extension schemas, see Resource Provider Schema in the CloudFormation CLI User Guide.

describeTypeResponse_publisherId :: Lens' DescribeTypeResponse (Maybe Text)

The publisher ID of the extension publisher.

This applies only to public third-party extensions. For private registered extensions, and extensions provided by Amazon, CloudFormation returns null.

describeTypeResponse_executionRoleArn :: Lens' DescribeTypeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM execution role used to register the extension. This applies only to private extensions you have registered in your account. For more information, see RegisterType.

If the registered extension calls any Amazon Web Services APIs, you must create an <https:docs.aws.amazon.comIAMlatestUserGuideid_roles.html IAM execution role> that includes the necessary permissions to call those Amazon Web Services APIs, and provision that execution role in your account. CloudFormation then assumes that execution role to provide your extension with the appropriate credentials.

describeTypeResponse_latestPublicVersion :: Lens' DescribeTypeResponse (Maybe Text)

The latest version of a public extension that is available for use.

This only applies if you specify a public extension, and you do not specify a version. For all other requests, CloudFormation returns null.

describeTypeResponse_arn :: Lens' DescribeTypeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the extension.

describeTypeResponse_deprecatedStatus :: Lens' DescribeTypeResponse (Maybe DeprecatedStatus)

The deprecation status of the extension version.

Valid values include:

LIVE: The extension is activated or registered and can be used in CloudFormation operations, dependent on its provisioning behavior and visibility scope.

DEPRECATED: The extension has been deactivated or deregistered and can no longer be used in CloudFormation operations.

For public third-party extensions, CloudFormation returns null.

describeTypeResponse_requiredActivatedTypes :: Lens' DescribeTypeResponse (Maybe [RequiredActivatedType])

For extensions that are modules, the public third-party extensions that must be activated in your account in order for the module itself to be activated.

describeTypeResponse_lastUpdated :: Lens' DescribeTypeResponse (Maybe UTCTime)

When the specified extension version was registered. This applies only to:

Private extensions you have registered in your account. For more information, see RegisterType.

Public extensions you have activated in your account with auto-update specified. For more information, see ActivateType.

describeTypeResponse_defaultVersionId :: Lens' DescribeTypeResponse (Maybe Text)

The ID of the default version of the extension. The default version is used when the extension version is not specified.

This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

To set the default version of an extension, use SetTypeDefaultVersion .

describeTypeResponse_documentationUrl :: Lens' DescribeTypeResponse (Maybe Text)

The URL of a page providing detailed documentation for this extension.

describeTypeResponse_provisioningType :: Lens' DescribeTypeResponse (Maybe ProvisioningType)

For resource type extensions, the provisioning behavior of the resource type. CloudFormation determines the provisioning type during registration, based on the types of handlers in the schema handler package submitted.

Valid values include:

FULLY_MUTABLE: The resource type includes an update handler to process updates to the type during stack update operations.

IMMUTABLE: The resource type does not include an update handler, so the type cannot be updated and must instead be replaced during stack update operations.

NON_PROVISIONABLE: The resource type does not include all of the following handlers, and therefore cannot actually be provisioned.

create

read

delete

describeTypeResponse_sourceUrl :: Lens' DescribeTypeResponse (Maybe Text)

The URL of the source code for the extension.

describeTypeResponse_description :: Lens' DescribeTypeResponse (Maybe Text)

The description of the extension.

describeTypeResponse_autoUpdate :: Lens' DescribeTypeResponse (Maybe Bool)

Whether CloudFormation automatically updates the extension in this account and region when a new minor version is published by the extension publisher. Major versions released by the publisher must be manually updated. For more information, see Activating public extensions for use in your account in the CloudFormation User Guide.

describeTypeResponse_originalTypeName :: Lens' DescribeTypeResponse (Maybe Text)

For public extensions that have been activated for this account and region, the type name of the public extension.

If you specified a TypeNameAlias when enabling the extension in this account and region, CloudFormation treats that alias as the extension's type name within the account and region, not the type name of the public extension. For more information, see Specifying aliases to refer to extensions in the CloudFormation User Guide.

describeTypeResponse_visibility :: Lens' DescribeTypeResponse (Maybe Visibility)

The scope at which the extension is visible and usable in CloudFormation operations.

Valid values include:

PRIVATE: The extension is only visible and usable within the account in which it is registered. CloudFormation marks any extensions you register as PRIVATE.

PUBLIC: The extension is publically visible and usable within any Amazon account.

describeTypeResponse_isDefaultVersion :: Lens' DescribeTypeResponse (Maybe Bool)

Whether the specified extension version is set as the default version.

This applies only to private extensions you have registered in your account, and extensions published by Amazon. For public third-party extensions, whether or not they are activated in your account, CloudFormation returns null.

describeTypeResponse_publicVersionNumber :: Lens' DescribeTypeResponse (Maybe Text)

The version number of a public third-party extension.

This applies only if you specify a public extension you have activated in your account, or specify a public extension without specifying a version. For all other extensions, CloudFormation returns null.

describeTypeResponse_type :: Lens' DescribeTypeResponse (Maybe RegistryType)

The kind of extension.

describeTypeResponse_timeCreated :: Lens' DescribeTypeResponse (Maybe UTCTime)

When the specified private extension version was registered or activated in your account.

describeTypeResponse_httpStatus :: Lens' DescribeTypeResponse Int

The response's http status code.