Network.AWS.CloudFormation.DescribeTypeRegistration

Description

Returns information about an extension's registration, including its current status and type and version identifiers.

When you initiate a registration request using RegisterType , you can then use DescribeTypeRegistration to monitor the progress of that registration request.

Once the registration request has completed, use DescribeType to return detailed information about an extension.

Creating a Request

data DescribeTypeRegistration

See: newDescribeTypeRegistration smart constructor.

Constructors

DescribeTypeRegistration'

Fields

registrationToken :: Text

The identifier for this registration request.

This registration token is generated by CloudFormation when you initiate a registration request using RegisterType .

newDescribeTypeRegistration

Arguments

:: Text
$sel:registrationToken:DescribeTypeRegistration'

-> DescribeTypeRegistration

Create a value of DescribeTypeRegistration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registrationToken:DescribeTypeRegistration', describeTypeRegistration_registrationToken - The identifier for this registration request.

This registration token is generated by CloudFormation when you initiate a registration request using RegisterType .

Request Lenses

describeTypeRegistration_registrationToken :: Lens' DescribeTypeRegistration Text

The identifier for this registration request.

This registration token is generated by CloudFormation when you initiate a registration request using RegisterType .

Destructuring the Response

data DescribeTypeRegistrationResponse

See: newDescribeTypeRegistrationResponse smart constructor.

Constructors

DescribeTypeRegistrationResponse'

Fields

typeVersionArn :: Maybe Text

The Amazon Resource Name (ARN) of this specific version of the extension being registered.

For registration requests with a ProgressStatus of other than COMPLETE, this will be null.

progressStatus :: Maybe RegistrationStatus

The current status of the extension registration request.

typeArn :: Maybe Text

The Amazon Resource Name (ARN) of the extension being registered.

For registration requests with a ProgressStatus of other than COMPLETE, this will be null.

description :: Maybe Text

The description of the extension registration request.

httpStatus :: Int

The response's http status code.

newDescribeTypeRegistrationResponse

Arguments

:: Int
$sel:httpStatus:DescribeTypeRegistrationResponse'

-> DescribeTypeRegistrationResponse

Create a value of DescribeTypeRegistrationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeVersionArn:DescribeTypeRegistrationResponse', describeTypeRegistrationResponse_typeVersionArn - The Amazon Resource Name (ARN) of this specific version of the extension being registered.

For registration requests with a ProgressStatus of other than COMPLETE, this will be null.

$sel:progressStatus:DescribeTypeRegistrationResponse', describeTypeRegistrationResponse_progressStatus - The current status of the extension registration request.

$sel:typeArn:DescribeTypeRegistrationResponse', describeTypeRegistrationResponse_typeArn - The Amazon Resource Name (ARN) of the extension being registered.

For registration requests with a ProgressStatus of other than COMPLETE, this will be null.

$sel:description:DescribeTypeRegistrationResponse', describeTypeRegistrationResponse_description - The description of the extension registration request.

$sel:httpStatus:DescribeTypeRegistrationResponse', describeTypeRegistrationResponse_httpStatus - The response's http status code.

Response Lenses

describeTypeRegistrationResponse_typeVersionArn :: Lens' DescribeTypeRegistrationResponse (Maybe Text)

The Amazon Resource Name (ARN) of this specific version of the extension being registered.

For registration requests with a ProgressStatus of other than COMPLETE, this will be null.

describeTypeRegistrationResponse_progressStatus :: Lens' DescribeTypeRegistrationResponse (Maybe RegistrationStatus)

The current status of the extension registration request.

describeTypeRegistrationResponse_typeArn :: Lens' DescribeTypeRegistrationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the extension being registered.

For registration requests with a ProgressStatus of other than COMPLETE, this will be null.

describeTypeRegistrationResponse_description :: Lens' DescribeTypeRegistrationResponse (Maybe Text)

The description of the extension registration request.

describeTypeRegistrationResponse_httpStatus :: Lens' DescribeTypeRegistrationResponse Int

The response's http status code.