Network.AWS.CloudFormation.DetectStackSetDrift

Description

Detect drift on a stack set. When CloudFormation performs drift detection on a stack set, it performs drift detection on the stack associated with each stack instance in the stack set. For more information, see How CloudFormation Performs Drift Detection on a Stack Set.

DetectStackSetDrift returns the OperationId of the stack set drift detection operation. Use this operation id with DescribeStackSetOperation to monitor the progress of the drift detection operation. The drift detection operation may take some time, depending on the number of stack instances included in the stack set, as well as the number of resources included in each stack.

Once the operation has completed, use the following actions to return drift information:

Use DescribeStackSet to return detailed information about the stack set, including detailed information about the last completed drift operation performed on the stack set. (Information about drift operations that are in progress is not included.)

Use ListStackInstances to return a list of stack instances belonging to the stack set, including the drift status and last drift time checked of each instance.

Use DescribeStackInstance to return detailed information about a specific stack instance, including its drift status and last drift time checked.

For more information on performing a drift detection operation on a stack set, see Detecting Unmanaged Changes in Stack Sets.

You can only run a single drift detection operation on a given stack set at one time.

To stop a drift detection stack set operation, use StopStackSetOperation .

Creating a Request

data DetectStackSetDrift

See: newDetectStackSetDrift smart constructor.

Constructors

DetectStackSetDrift'

Fields

operationId :: Maybe Text

The ID of the stack set operation.

callAs :: Maybe CallAs

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

operationPreferences :: Maybe StackSetOperationPreferences

stackSetName :: Text

The name of the stack set on which to perform the drift detection operation.

newDetectStackSetDrift

Arguments

:: Text
$sel:stackSetName:DetectStackSetDrift'

-> DetectStackSetDrift

Create a value of DetectStackSetDrift with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:DetectStackSetDrift', detectStackSetDrift_operationId - The ID of the stack set operation.

$sel:callAs:DetectStackSetDrift', detectStackSetDrift_callAs - [Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

$sel:operationPreferences:DetectStackSetDrift', detectStackSetDrift_operationPreferences - Undocumented member.

$sel:stackSetName:DetectStackSetDrift', detectStackSetDrift_stackSetName - The name of the stack set on which to perform the drift detection operation.

Request Lenses

detectStackSetDrift_operationId :: Lens' DetectStackSetDrift (Maybe Text)

The ID of the stack set operation.

detectStackSetDrift_callAs :: Lens' DetectStackSetDrift (Maybe CallAs)

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

detectStackSetDrift_operationPreferences :: Lens' DetectStackSetDrift (Maybe StackSetOperationPreferences)

Undocumented member.

detectStackSetDrift_stackSetName :: Lens' DetectStackSetDrift Text

The name of the stack set on which to perform the drift detection operation.

Destructuring the Response

data DetectStackSetDriftResponse

See: newDetectStackSetDriftResponse smart constructor.

Constructors

DetectStackSetDriftResponse'

Fields

operationId :: Maybe Text

The ID of the drift detection stack set operation.

you can use this operation id with DescribeStackSetOperation to monitor the progress of the drift detection operation.

httpStatus :: Int

The response's http status code.

newDetectStackSetDriftResponse

Arguments

:: Int
$sel:httpStatus:DetectStackSetDriftResponse'

-> DetectStackSetDriftResponse

Create a value of DetectStackSetDriftResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:DetectStackSetDrift', detectStackSetDriftResponse_operationId - The ID of the drift detection stack set operation.

you can use this operation id with DescribeStackSetOperation to monitor the progress of the drift detection operation.

$sel:httpStatus:DetectStackSetDriftResponse', detectStackSetDriftResponse_httpStatus - The response's http status code.

Response Lenses

detectStackSetDriftResponse_operationId :: Lens' DetectStackSetDriftResponse (Maybe Text)

The ID of the drift detection stack set operation.

you can use this operation id with DescribeStackSetOperation to monitor the progress of the drift detection operation.

detectStackSetDriftResponse_httpStatus :: Lens' DetectStackSetDriftResponse Int

The response's http status code.