Network.AWS.CloudFormation.EstimateTemplateCost

Description

Returns the estimated monthly cost of a template. The return value is an Amazon Web Services Simple Monthly Calculator URL with a query string that describes the resources required to run the template.

Creating a Request

data EstimateTemplateCost

The input for an EstimateTemplateCost action.

See: newEstimateTemplateCost smart constructor.

Constructors

EstimateTemplateCost'

Fields

templateURL :: Maybe Text

Location of file containing the template body. The URL must point to a template that is located in an Amazon S3 bucket or a Systems Manager document. For more information, go to Template Anatomy in the CloudFormation User Guide.

Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

templateBody :: Maybe Text

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to Template Anatomy in the CloudFormation User Guide.)

Conditional: You must pass TemplateBody or TemplateURL. If both are passed, only TemplateBody is used.

parameters :: Maybe [Parameter]

A list of Parameter structures that specify input parameters.

newEstimateTemplateCost :: EstimateTemplateCost

Create a value of EstimateTemplateCost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateURL:EstimateTemplateCost', estimateTemplateCost_templateURL - Location of file containing the template body. The URL must point to a template that is located in an Amazon S3 bucket or a Systems Manager document. For more information, go to Template Anatomy in the CloudFormation User Guide.

Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

$sel:templateBody:EstimateTemplateCost', estimateTemplateCost_templateBody - Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to Template Anatomy in the CloudFormation User Guide.)

Conditional: You must pass TemplateBody or TemplateURL. If both are passed, only TemplateBody is used.

$sel:parameters:EstimateTemplateCost', estimateTemplateCost_parameters - A list of Parameter structures that specify input parameters.

Request Lenses

estimateTemplateCost_templateURL :: Lens' EstimateTemplateCost (Maybe Text)

Location of file containing the template body. The URL must point to a template that is located in an Amazon S3 bucket or a Systems Manager document. For more information, go to Template Anatomy in the CloudFormation User Guide.

Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

estimateTemplateCost_templateBody :: Lens' EstimateTemplateCost (Maybe Text)

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to Template Anatomy in the CloudFormation User Guide.)

Conditional: You must pass TemplateBody or TemplateURL. If both are passed, only TemplateBody is used.

estimateTemplateCost_parameters :: Lens' EstimateTemplateCost (Maybe [Parameter])

A list of Parameter structures that specify input parameters.

Destructuring the Response

data EstimateTemplateCostResponse

The output for a EstimateTemplateCost action.

See: newEstimateTemplateCostResponse smart constructor.

Constructors

EstimateTemplateCostResponse'

Fields

url :: Maybe Text

An Amazon Web Services Simple Monthly Calculator URL with a query string that describes the resources required to run the template.

httpStatus :: Int

The response's http status code.

newEstimateTemplateCostResponse

Arguments

:: Int
$sel:httpStatus:EstimateTemplateCostResponse'

-> EstimateTemplateCostResponse

Create a value of EstimateTemplateCostResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:EstimateTemplateCostResponse', estimateTemplateCostResponse_url - An Amazon Web Services Simple Monthly Calculator URL with a query string that describes the resources required to run the template.

$sel:httpStatus:EstimateTemplateCostResponse', estimateTemplateCostResponse_httpStatus - The response's http status code.

Response Lenses

estimateTemplateCostResponse_url :: Lens' EstimateTemplateCostResponse (Maybe Text)

An Amazon Web Services Simple Monthly Calculator URL with a query string that describes the resources required to run the template.

estimateTemplateCostResponse_httpStatus :: Lens' EstimateTemplateCostResponse Int

The response's http status code.