Network.AWS.CloudFormation.GetTemplateSummary

Description

Returns information about a new or existing template. The GetTemplateSummary action is useful for viewing parameter information, such as default parameter values and parameter types, before you create or update a stack or stack set.

You can use the GetTemplateSummary action when you submit a template, or you can get template information for a stack set, or a running or deleted stack.

For deleted stacks, GetTemplateSummary returns the template information for up to 90 days after the stack has been deleted. If the template does not exist, a ValidationError is returned.

Creating a Request

data GetTemplateSummary

The input for the GetTemplateSummary action.

See: newGetTemplateSummary smart constructor.

Constructors

GetTemplateSummary'

Fields

stackName :: Maybe Text

The name or the stack ID that is associated with the stack, which are not always interchangeable. For running stacks, you can specify either the stack's name or its unique stack ID. For deleted stack, you must specify the unique stack ID.

Conditional: You must specify only one of the following parameters: StackName, StackSetName, TemplateBody, or TemplateURL.

templateURL :: Maybe Text

Location of file containing the template body. The URL must point to a template (max size: 460,800 bytes) that is located in an Amazon S3 bucket or a Systems Manager document. For more information about templates, see Template Anatomy in the CloudFormation User Guide.

Conditional: You must specify only one of the following parameters: StackName, StackSetName, TemplateBody, or TemplateURL.

callAs :: Maybe CallAs

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

stackSetName :: Maybe Text

The name or unique ID of the stack set from which the stack was created.

Conditional: You must specify only one of the following parameters: StackName, StackSetName, TemplateBody, or TemplateURL.

templateBody :: Maybe Text

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information about templates, see Template Anatomy in the CloudFormation User Guide.

Conditional: You must specify only one of the following parameters: StackName, StackSetName, TemplateBody, or TemplateURL.

newGetTemplateSummary :: GetTemplateSummary

Create a value of GetTemplateSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackName:GetTemplateSummary', getTemplateSummary_stackName - The name or the stack ID that is associated with the stack, which are not always interchangeable. For running stacks, you can specify either the stack's name or its unique stack ID. For deleted stack, you must specify the unique stack ID.

Conditional: You must specify only one of the following parameters: StackName, StackSetName, TemplateBody, or TemplateURL.

$sel:templateURL:GetTemplateSummary', getTemplateSummary_templateURL - Location of file containing the template body. The URL must point to a template (max size: 460,800 bytes) that is located in an Amazon S3 bucket or a Systems Manager document. For more information about templates, see Template Anatomy in the CloudFormation User Guide.

Conditional: You must specify only one of the following parameters: StackName, StackSetName, TemplateBody, or TemplateURL.

$sel:callAs:GetTemplateSummary', getTemplateSummary_callAs - [Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

$sel:stackSetName:GetTemplateSummary', getTemplateSummary_stackSetName - The name or unique ID of the stack set from which the stack was created.

Conditional: You must specify only one of the following parameters: StackName, StackSetName, TemplateBody, or TemplateURL.

$sel:templateBody:GetTemplateSummary', getTemplateSummary_templateBody - Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information about templates, see Template Anatomy in the CloudFormation User Guide.

Conditional: You must specify only one of the following parameters: StackName, StackSetName, TemplateBody, or TemplateURL.

Request Lenses

getTemplateSummary_stackName :: Lens' GetTemplateSummary (Maybe Text)

The name or the stack ID that is associated with the stack, which are not always interchangeable. For running stacks, you can specify either the stack's name or its unique stack ID. For deleted stack, you must specify the unique stack ID.

Conditional: You must specify only one of the following parameters: StackName, StackSetName, TemplateBody, or TemplateURL.

getTemplateSummary_templateURL :: Lens' GetTemplateSummary (Maybe Text)

Location of file containing the template body. The URL must point to a template (max size: 460,800 bytes) that is located in an Amazon S3 bucket or a Systems Manager document. For more information about templates, see Template Anatomy in the CloudFormation User Guide.

Conditional: You must specify only one of the following parameters: StackName, StackSetName, TemplateBody, or TemplateURL.

getTemplateSummary_callAs :: Lens' GetTemplateSummary (Maybe CallAs)

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

getTemplateSummary_stackSetName :: Lens' GetTemplateSummary (Maybe Text)

The name or unique ID of the stack set from which the stack was created.

Conditional: You must specify only one of the following parameters: StackName, StackSetName, TemplateBody, or TemplateURL.

getTemplateSummary_templateBody :: Lens' GetTemplateSummary (Maybe Text)

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information about templates, see Template Anatomy in the CloudFormation User Guide.

Conditional: You must specify only one of the following parameters: StackName, StackSetName, TemplateBody, or TemplateURL.

Destructuring the Response

data GetTemplateSummaryResponse

The output for the GetTemplateSummary action.

See: newGetTemplateSummaryResponse smart constructor.

Constructors

GetTemplateSummaryResponse'

Fields

resourceTypes :: Maybe [Text]

A list of all the template resource types that are defined in the template, such as AWS::EC2::Instance, AWS::Dynamo::Table, and Custom::MyCustomInstance.

capabilities :: Maybe [Capability]

The capabilities found within the template. If your template contains IAM resources, you must specify the CAPABILITY_IAM or CAPABILITY_NAMED_IAM value for this parameter when you use the CreateStack or UpdateStack actions with your template; otherwise, those actions return an InsufficientCapabilities error.

For more information, see Acknowledging IAM Resources in CloudFormation Templates.

resourceIdentifierSummaries :: Maybe [ResourceIdentifierSummary]

A list of resource identifier summaries that describe the target resources of an import operation and the properties you can provide during the import to identify the target resources. For example, BucketName is a possible identifier property for an AWS::S3::Bucket resource.

declaredTransforms :: Maybe [Text]

A list of the transforms that are declared in the template.

metadata :: Maybe Text

The value that is defined for the Metadata property of the template.

version :: Maybe Text

The Amazon Web Services template format version, which identifies the capabilities of the template.

description :: Maybe Text

The value that is defined in the Description property of the template.

parameters :: Maybe [ParameterDeclaration]

A list of parameter declarations that describe various properties for each parameter.

capabilitiesReason :: Maybe Text

The list of resources that generated the values in the Capabilities response element.

httpStatus :: Int

The response's http status code.

newGetTemplateSummaryResponse

Arguments

:: Int
$sel:httpStatus:GetTemplateSummaryResponse'

-> GetTemplateSummaryResponse

Create a value of GetTemplateSummaryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceTypes:GetTemplateSummaryResponse', getTemplateSummaryResponse_resourceTypes - A list of all the template resource types that are defined in the template, such as AWS::EC2::Instance, AWS::Dynamo::Table, and Custom::MyCustomInstance.

$sel:capabilities:GetTemplateSummaryResponse', getTemplateSummaryResponse_capabilities - The capabilities found within the template. If your template contains IAM resources, you must specify the CAPABILITY_IAM or CAPABILITY_NAMED_IAM value for this parameter when you use the CreateStack or UpdateStack actions with your template; otherwise, those actions return an InsufficientCapabilities error.

For more information, see Acknowledging IAM Resources in CloudFormation Templates.

$sel:resourceIdentifierSummaries:GetTemplateSummaryResponse', getTemplateSummaryResponse_resourceIdentifierSummaries - A list of resource identifier summaries that describe the target resources of an import operation and the properties you can provide during the import to identify the target resources. For example, BucketName is a possible identifier property for an AWS::S3::Bucket resource.

$sel:declaredTransforms:GetTemplateSummaryResponse', getTemplateSummaryResponse_declaredTransforms - A list of the transforms that are declared in the template.

$sel:metadata:GetTemplateSummaryResponse', getTemplateSummaryResponse_metadata - The value that is defined for the Metadata property of the template.

$sel:version:GetTemplateSummaryResponse', getTemplateSummaryResponse_version - The Amazon Web Services template format version, which identifies the capabilities of the template.

$sel:description:GetTemplateSummaryResponse', getTemplateSummaryResponse_description - The value that is defined in the Description property of the template.

$sel:parameters:GetTemplateSummaryResponse', getTemplateSummaryResponse_parameters - A list of parameter declarations that describe various properties for each parameter.

$sel:capabilitiesReason:GetTemplateSummaryResponse', getTemplateSummaryResponse_capabilitiesReason - The list of resources that generated the values in the Capabilities response element.

$sel:httpStatus:GetTemplateSummaryResponse', getTemplateSummaryResponse_httpStatus - The response's http status code.

Response Lenses

getTemplateSummaryResponse_resourceTypes :: Lens' GetTemplateSummaryResponse (Maybe [Text])

A list of all the template resource types that are defined in the template, such as AWS::EC2::Instance, AWS::Dynamo::Table, and Custom::MyCustomInstance.

getTemplateSummaryResponse_capabilities :: Lens' GetTemplateSummaryResponse (Maybe [Capability])

The capabilities found within the template. If your template contains IAM resources, you must specify the CAPABILITY_IAM or CAPABILITY_NAMED_IAM value for this parameter when you use the CreateStack or UpdateStack actions with your template; otherwise, those actions return an InsufficientCapabilities error.

For more information, see Acknowledging IAM Resources in CloudFormation Templates.

getTemplateSummaryResponse_resourceIdentifierSummaries :: Lens' GetTemplateSummaryResponse (Maybe [ResourceIdentifierSummary])

A list of resource identifier summaries that describe the target resources of an import operation and the properties you can provide during the import to identify the target resources. For example, BucketName is a possible identifier property for an AWS::S3::Bucket resource.

getTemplateSummaryResponse_declaredTransforms :: Lens' GetTemplateSummaryResponse (Maybe [Text])

A list of the transforms that are declared in the template.

getTemplateSummaryResponse_metadata :: Lens' GetTemplateSummaryResponse (Maybe Text)

The value that is defined for the Metadata property of the template.

getTemplateSummaryResponse_version :: Lens' GetTemplateSummaryResponse (Maybe Text)

The Amazon Web Services template format version, which identifies the capabilities of the template.

getTemplateSummaryResponse_description :: Lens' GetTemplateSummaryResponse (Maybe Text)

The value that is defined in the Description property of the template.

getTemplateSummaryResponse_parameters :: Lens' GetTemplateSummaryResponse (Maybe [ParameterDeclaration])

A list of parameter declarations that describe various properties for each parameter.

getTemplateSummaryResponse_capabilitiesReason :: Lens' GetTemplateSummaryResponse (Maybe Text)

The list of resources that generated the values in the Capabilities response element.

getTemplateSummaryResponse_httpStatus :: Lens' GetTemplateSummaryResponse Int

The response's http status code.