Network.AWS.CloudFormation.ImportStacksToStackSet

Description

Import existing stacks into a new stack sets. Use the stack import operation to import up to 10 stacks into a new stack set in the same account as the source stack or in a different administrator account and Region, by specifying the stack ID of the stack you intend to import.

ImportStacksToStackSet is only supported by self-managed permissions.

Creating a Request

data ImportStacksToStackSet

See: newImportStacksToStackSet smart constructor.

Constructors

ImportStacksToStackSet'

Fields

operationId :: Maybe Text

A unique, user defined, identifier for the stack set operation.

callAs :: Maybe CallAs

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

For service managed stack sets, specify DELEGATED_ADMIN.

operationPreferences :: Maybe StackSetOperationPreferences

stackSetName :: Text

The name of the stack set. The name must be unique in the Region where you create your stack set.

stackIds :: [Text]

The IDs of the stacks you are importing into a stack set. You import up to 10 stacks per stack set at a time.

newImportStacksToStackSet

Arguments

:: Text
$sel:stackSetName:ImportStacksToStackSet'

-> ImportStacksToStackSet

Create a value of ImportStacksToStackSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:ImportStacksToStackSet', importStacksToStackSet_operationId - A unique, user defined, identifier for the stack set operation.

$sel:callAs:ImportStacksToStackSet', importStacksToStackSet_callAs - By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

For service managed stack sets, specify DELEGATED_ADMIN.

$sel:operationPreferences:ImportStacksToStackSet', importStacksToStackSet_operationPreferences - Undocumented member.

$sel:stackSetName:ImportStacksToStackSet', importStacksToStackSet_stackSetName - The name of the stack set. The name must be unique in the Region where you create your stack set.

$sel:stackIds:ImportStacksToStackSet', importStacksToStackSet_stackIds - The IDs of the stacks you are importing into a stack set. You import up to 10 stacks per stack set at a time.

Request Lenses

importStacksToStackSet_operationId :: Lens' ImportStacksToStackSet (Maybe Text)

A unique, user defined, identifier for the stack set operation.

importStacksToStackSet_callAs :: Lens' ImportStacksToStackSet (Maybe CallAs)

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

For service managed stack sets, specify DELEGATED_ADMIN.

importStacksToStackSet_operationPreferences :: Lens' ImportStacksToStackSet (Maybe StackSetOperationPreferences)

Undocumented member.

importStacksToStackSet_stackSetName :: Lens' ImportStacksToStackSet Text

The name of the stack set. The name must be unique in the Region where you create your stack set.

importStacksToStackSet_stackIds :: Lens' ImportStacksToStackSet [Text]

The IDs of the stacks you are importing into a stack set. You import up to 10 stacks per stack set at a time.

Destructuring the Response

data ImportStacksToStackSetResponse

See: newImportStacksToStackSetResponse smart constructor.

Constructors

ImportStacksToStackSetResponse'

Fields

operationId :: Maybe Text

The unique identifier for the stack set operation.

httpStatus :: Int

The response's http status code.

newImportStacksToStackSetResponse

Arguments

:: Int
$sel:httpStatus:ImportStacksToStackSetResponse'

-> ImportStacksToStackSetResponse

Create a value of ImportStacksToStackSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:ImportStacksToStackSet', importStacksToStackSetResponse_operationId - The unique identifier for the stack set operation.

$sel:httpStatus:ImportStacksToStackSetResponse', importStacksToStackSetResponse_httpStatus - The response's http status code.

Response Lenses

importStacksToStackSetResponse_operationId :: Lens' ImportStacksToStackSetResponse (Maybe Text)

The unique identifier for the stack set operation.

importStacksToStackSetResponse_httpStatus :: Lens' ImportStacksToStackSetResponse Int

The response's http status code.