Network.AWS.CloudFormation.ListStackInstances

Description

Returns summary information about stack instances that are associated with the specified stack set. You can filter for stack instances that are associated with a specific Amazon Web Services account name or Region, or that have a specific status.

This operation returns paginated results.

Creating a Request

data ListStackInstances

See: newListStackInstances smart constructor.

Constructors

ListStackInstances'

Fields

nextToken :: Maybe Text

If the previous request didn't return all of the remaining results, the response's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

stackInstanceAccount :: Maybe Text

The name of the Amazon Web Services account that you want to list stack instances for.

maxResults :: Maybe Natural

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

callAs :: Maybe CallAs

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

stackInstanceRegion :: Maybe Text

The name of the Region where you want to list stack instances.

filters :: Maybe [StackInstanceFilter]

The status that stack instances are filtered by.

stackSetName :: Text

The name or unique ID of the stack set that you want to list stack instances for.

newListStackInstances

Arguments

:: Text
$sel:stackSetName:ListStackInstances'

-> ListStackInstances

Create a value of ListStackInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListStackInstances', listStackInstances_nextToken - If the previous request didn't return all of the remaining results, the response's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:stackInstanceAccount:ListStackInstances', listStackInstances_stackInstanceAccount - The name of the Amazon Web Services account that you want to list stack instances for.

$sel:maxResults:ListStackInstances', listStackInstances_maxResults - The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

$sel:callAs:ListStackInstances', listStackInstances_callAs - [Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

$sel:stackInstanceRegion:ListStackInstances', listStackInstances_stackInstanceRegion - The name of the Region where you want to list stack instances.

$sel:filters:ListStackInstances', listStackInstances_filters - The status that stack instances are filtered by.

$sel:stackSetName:ListStackInstances', listStackInstances_stackSetName - The name or unique ID of the stack set that you want to list stack instances for.

Request Lenses

listStackInstances_nextToken :: Lens' ListStackInstances (Maybe Text)

If the previous request didn't return all of the remaining results, the response's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listStackInstances_stackInstanceAccount :: Lens' ListStackInstances (Maybe Text)

The name of the Amazon Web Services account that you want to list stack instances for.

listStackInstances_maxResults :: Lens' ListStackInstances (Maybe Natural)

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

listStackInstances_callAs :: Lens' ListStackInstances (Maybe CallAs)

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

listStackInstances_stackInstanceRegion :: Lens' ListStackInstances (Maybe Text)

The name of the Region where you want to list stack instances.

listStackInstances_filters :: Lens' ListStackInstances (Maybe [StackInstanceFilter])

The status that stack instances are filtered by.

listStackInstances_stackSetName :: Lens' ListStackInstances Text

The name or unique ID of the stack set that you want to list stack instances for.

Destructuring the Response

data ListStackInstancesResponse

See: newListStackInstancesResponse smart constructor.

Constructors

ListStackInstancesResponse'

Fields

nextToken :: Maybe Text

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

summaries :: Maybe [StackInstanceSummary]

A list of StackInstanceSummary structures that contain information about the specified stack instances.

httpStatus :: Int

The response's http status code.

newListStackInstancesResponse

Arguments

:: Int
$sel:httpStatus:ListStackInstancesResponse'

-> ListStackInstancesResponse

Create a value of ListStackInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListStackInstances', listStackInstancesResponse_nextToken - If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

$sel:summaries:ListStackInstancesResponse', listStackInstancesResponse_summaries - A list of StackInstanceSummary structures that contain information about the specified stack instances.

$sel:httpStatus:ListStackInstancesResponse', listStackInstancesResponse_httpStatus - The response's http status code.

Response Lenses

listStackInstancesResponse_nextToken :: Lens' ListStackInstancesResponse (Maybe Text)

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

listStackInstancesResponse_summaries :: Lens' ListStackInstancesResponse (Maybe [StackInstanceSummary])

A list of StackInstanceSummary structures that contain information about the specified stack instances.

listStackInstancesResponse_httpStatus :: Lens' ListStackInstancesResponse Int

The response's http status code.