Network.AWS.CloudFormation.ListStackSetOperationResults

Description

Returns summary information about the results of a stack set operation.

This operation returns paginated results.

Creating a Request

data ListStackSetOperationResults

See: newListStackSetOperationResults smart constructor.

Constructors

ListStackSetOperationResults'

Fields

nextToken :: Maybe Text

If the previous request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackSetOperationResults again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

maxResults :: Maybe Natural

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

callAs :: Maybe CallAs

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

stackSetName :: Text

The name or unique ID of the stack set that you want to get operation results for.

operationId :: Text

The ID of the stack set operation.

newListStackSetOperationResults

Arguments

:: Text
$sel:stackSetName:ListStackSetOperationResults'

-> Text
$sel:operationId:ListStackSetOperationResults'

-> ListStackSetOperationResults

Create a value of ListStackSetOperationResults with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListStackSetOperationResults', listStackSetOperationResults_nextToken - If the previous request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackSetOperationResults again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:maxResults:ListStackSetOperationResults', listStackSetOperationResults_maxResults - The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

$sel:callAs:ListStackSetOperationResults', listStackSetOperationResults_callAs - [Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

$sel:stackSetName:ListStackSetOperationResults', listStackSetOperationResults_stackSetName - The name or unique ID of the stack set that you want to get operation results for.

$sel:operationId:ListStackSetOperationResults', listStackSetOperationResults_operationId - The ID of the stack set operation.

Request Lenses

listStackSetOperationResults_nextToken :: Lens' ListStackSetOperationResults (Maybe Text)

If the previous request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackSetOperationResults again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listStackSetOperationResults_maxResults :: Lens' ListStackSetOperationResults (Maybe Natural)

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

listStackSetOperationResults_callAs :: Lens' ListStackSetOperationResults (Maybe CallAs)

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

listStackSetOperationResults_stackSetName :: Lens' ListStackSetOperationResults Text

The name or unique ID of the stack set that you want to get operation results for.

listStackSetOperationResults_operationId :: Lens' ListStackSetOperationResults Text

The ID of the stack set operation.

Destructuring the Response

data ListStackSetOperationResultsResponse

See: newListStackSetOperationResultsResponse smart constructor.

Constructors

ListStackSetOperationResultsResponse'

Fields

nextToken :: Maybe Text

If the request doesn't return all results, NextToken is set to a token. To retrieve the next set of results, call ListOperationResults again and assign that token to the request object's NextToken parameter. If there are no remaining results, NextToken is set to null.

summaries :: Maybe [StackSetOperationResultSummary]

A list of StackSetOperationResultSummary structures that contain information about the specified operation results, for accounts and Regions that are included in the operation.

httpStatus :: Int

The response's http status code.

newListStackSetOperationResultsResponse

Arguments

:: Int
$sel:httpStatus:ListStackSetOperationResultsResponse'

-> ListStackSetOperationResultsResponse

Create a value of ListStackSetOperationResultsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListStackSetOperationResults', listStackSetOperationResultsResponse_nextToken - If the request doesn't return all results, NextToken is set to a token. To retrieve the next set of results, call ListOperationResults again and assign that token to the request object's NextToken parameter. If there are no remaining results, NextToken is set to null.

$sel:summaries:ListStackSetOperationResultsResponse', listStackSetOperationResultsResponse_summaries - A list of StackSetOperationResultSummary structures that contain information about the specified operation results, for accounts and Regions that are included in the operation.

$sel:httpStatus:ListStackSetOperationResultsResponse', listStackSetOperationResultsResponse_httpStatus - The response's http status code.

Response Lenses

listStackSetOperationResultsResponse_nextToken :: Lens' ListStackSetOperationResultsResponse (Maybe Text)

If the request doesn't return all results, NextToken is set to a token. To retrieve the next set of results, call ListOperationResults again and assign that token to the request object's NextToken parameter. If there are no remaining results, NextToken is set to null.

listStackSetOperationResultsResponse_summaries :: Lens' ListStackSetOperationResultsResponse (Maybe [StackSetOperationResultSummary])

A list of StackSetOperationResultSummary structures that contain information about the specified operation results, for accounts and Regions that are included in the operation.

listStackSetOperationResultsResponse_httpStatus :: Lens' ListStackSetOperationResultsResponse Int

The response's http status code.