Network.AWS.CloudFormation.ListStackSets

Description

Returns summary information about stack sets that are associated with the user.

[Self-managed permissions] If you set the CallAs parameter to SELF while signed in to your Amazon Web Services account, ListStackSets returns all self-managed stack sets in your Amazon Web Services account.

[Service-managed permissions] If you set the CallAs parameter to SELF while signed in to the organization's management account, ListStackSets returns all stack sets in the management account.

[Service-managed permissions] If you set the CallAs parameter to DELEGATED_ADMIN while signed in to your member account, ListStackSets returns all stack sets with service-managed permissions in the management account.

This operation returns paginated results.

Creating a Request

data ListStackSets

See: newListStackSets smart constructor.

Constructors

ListStackSets'

Fields

status :: Maybe StackSetStatus

The status of the stack sets that you want to get summary information about.

callAs :: Maybe CallAs

Service-managed permissions

Specifies whether you are acting as an account administrator in the management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

nextToken :: Maybe Text

If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackSets again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

maxResults :: Maybe Natural

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

newListStackSets :: ListStackSets

Create a value of ListStackSets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListStackSets', listStackSets_status - The status of the stack sets that you want to get summary information about.

$sel:callAs:ListStackSets', listStackSets_callAs - [Service-managed permissions] Specifies whether you are acting as an account administrator in the management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

$sel:nextToken:ListStackSets', listStackSets_nextToken - If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackSets again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:maxResults:ListStackSets', listStackSets_maxResults - The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

Request Lenses

listStackSets_status :: Lens' ListStackSets (Maybe StackSetStatus)

The status of the stack sets that you want to get summary information about.

listStackSets_callAs :: Lens' ListStackSets (Maybe CallAs)

Service-managed permissions

Specifies whether you are acting as an account administrator in the management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

listStackSets_nextToken :: Lens' ListStackSets (Maybe Text)

If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackSets again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listStackSets_maxResults :: Lens' ListStackSets (Maybe Natural)

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

Destructuring the Response

data ListStackSetsResponse

See: newListStackSetsResponse smart constructor.

Constructors

ListStackSetsResponse'

Fields

nextToken :: Maybe Text

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

summaries :: Maybe [StackSetSummary]

A list of StackSetSummary structures that contain information about the user's stack sets.

httpStatus :: Int

The response's http status code.

newListStackSetsResponse

Arguments

:: Int
$sel:httpStatus:ListStackSetsResponse'

-> ListStackSetsResponse

Create a value of ListStackSetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListStackSets', listStackSetsResponse_nextToken - If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

$sel:summaries:ListStackSetsResponse', listStackSetsResponse_summaries - A list of StackSetSummary structures that contain information about the user's stack sets.

$sel:httpStatus:ListStackSetsResponse', listStackSetsResponse_httpStatus - The response's http status code.

Response Lenses

listStackSetsResponse_nextToken :: Lens' ListStackSetsResponse (Maybe Text)

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

listStackSetsResponse_summaries :: Lens' ListStackSetsResponse (Maybe [StackSetSummary])

A list of StackSetSummary structures that contain information about the user's stack sets.

listStackSetsResponse_httpStatus :: Lens' ListStackSetsResponse Int

The response's http status code.