Network.AWS.CloudFormation.ListStacks

Description

Returns the summary information for stacks whose status matches the specified StackStatusFilter. Summary information for stacks that have been deleted is kept for 90 days after the stack is deleted. If no StackStatusFilter is specified, summary information for all stacks is returned (including existing stacks and stacks that have been deleted).

This operation returns paginated results.

Creating a Request

data ListStacks

The input for ListStacks action.

See: newListStacks smart constructor.

Constructors

ListStacks'

Fields

nextToken :: Maybe Text

A string that identifies the next page of stacks that you want to retrieve.

stackStatusFilter :: Maybe [StackStatus]

Stack status to use as a filter. Specify one or more stack status codes to list only stacks with the specified status codes. For a complete list of stack status codes, see the StackStatus parameter of the Stack data type.

newListStacks :: ListStacks

Create a value of ListStacks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListStacks', listStacks_nextToken - A string that identifies the next page of stacks that you want to retrieve.

$sel:stackStatusFilter:ListStacks', listStacks_stackStatusFilter - Stack status to use as a filter. Specify one or more stack status codes to list only stacks with the specified status codes. For a complete list of stack status codes, see the StackStatus parameter of the Stack data type.

Request Lenses

listStacks_nextToken :: Lens' ListStacks (Maybe Text)

A string that identifies the next page of stacks that you want to retrieve.

listStacks_stackStatusFilter :: Lens' ListStacks (Maybe [StackStatus])

Stack status to use as a filter. Specify one or more stack status codes to list only stacks with the specified status codes. For a complete list of stack status codes, see the StackStatus parameter of the Stack data type.

Destructuring the Response

data ListStacksResponse

The output for ListStacks action.

See: newListStacksResponse smart constructor.

Constructors

ListStacksResponse'

Fields

nextToken :: Maybe Text

If the output exceeds 1 MB in size, a string that identifies the next page of stacks. If no additional page exists, this value is null.

stackSummaries :: Maybe [StackSummary]

A list of StackSummary structures containing information about the specified stacks.

httpStatus :: Int

The response's http status code.

newListStacksResponse

Arguments

:: Int
$sel:httpStatus:ListStacksResponse'

-> ListStacksResponse

Create a value of ListStacksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListStacks', listStacksResponse_nextToken - If the output exceeds 1 MB in size, a string that identifies the next page of stacks. If no additional page exists, this value is null.

$sel:stackSummaries:ListStacksResponse', listStacksResponse_stackSummaries - A list of StackSummary structures containing information about the specified stacks.

$sel:httpStatus:ListStacksResponse', listStacksResponse_httpStatus - The response's http status code.

Response Lenses

listStacksResponse_nextToken :: Lens' ListStacksResponse (Maybe Text)

If the output exceeds 1 MB in size, a string that identifies the next page of stacks. If no additional page exists, this value is null.

listStacksResponse_stackSummaries :: Lens' ListStacksResponse (Maybe [StackSummary])

A list of StackSummary structures containing information about the specified stacks.

listStacksResponse_httpStatus :: Lens' ListStacksResponse Int

The response's http status code.