Network.AWS.CloudFormation.ListTypes

Description

Returns summary information about extension that have been registered with CloudFormation.

This operation returns paginated results.

Creating a Request

data ListTypes

See: newListTypes smart constructor.

Constructors

ListTypes'

Fields

nextToken :: Maybe Text

If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

maxResults :: Maybe Natural

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

deprecatedStatus :: Maybe DeprecatedStatus

The deprecation status of the extension that you want to get summary information about.

Valid values include:

LIVE: The extension is registered for use in CloudFormation operations.

DEPRECATED: The extension has been deregistered and can no longer be used in CloudFormation operations.

provisioningType :: Maybe ProvisioningType

For resource types, the provisioning behavior of the resource type. CloudFormation determines the provisioning type during registration, based on the types of handlers in the schema handler package submitted.

Valid values include:

FULLY_MUTABLE: The resource type includes an update handler to process updates to the type during stack update operations.

IMMUTABLE: The resource type does not include an update handler, so the type cannot be updated and must instead be replaced during stack update operations.

NON_PROVISIONABLE: The resource type does not include create, read, and delete handlers, and therefore cannot actually be provisioned.

The default is FULLY_MUTABLE.

visibility :: Maybe Visibility

The scope at which the extensions are visible and usable in CloudFormation operations.

Valid values include:

PRIVATE: Extensions that are visible and usable within this account and region. This includes:

Private extensions you have registered in this account and region.

Public extensions that you have activated in this account and region.

PUBLIC: Extensions that are publicly visible and available to be activated within any Amazon account. This includes extensions from Amazon, as well as third-party publishers.

The default is PRIVATE.

filters :: Maybe TypeFilters

Filter criteria to use in determining which extensions to return.

If you specify a filter, CloudFormation ignores any specified Visibility value when returning the list of types.

type' :: Maybe RegistryType

The type of extension.

newListTypes :: ListTypes

Create a value of ListTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTypes', listTypes_nextToken - If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

$sel:maxResults:ListTypes', listTypes_maxResults - The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

$sel:deprecatedStatus:ListTypes', listTypes_deprecatedStatus - The deprecation status of the extension that you want to get summary information about.

Valid values include:

LIVE: The extension is registered for use in CloudFormation operations.

DEPRECATED: The extension has been deregistered and can no longer be used in CloudFormation operations.

$sel:provisioningType:ListTypes', listTypes_provisioningType - For resource types, the provisioning behavior of the resource type. CloudFormation determines the provisioning type during registration, based on the types of handlers in the schema handler package submitted.

Valid values include:

FULLY_MUTABLE: The resource type includes an update handler to process updates to the type during stack update operations.

IMMUTABLE: The resource type does not include an update handler, so the type cannot be updated and must instead be replaced during stack update operations.

NON_PROVISIONABLE: The resource type does not include create, read, and delete handlers, and therefore cannot actually be provisioned.

The default is FULLY_MUTABLE.

$sel:visibility:ListTypes', listTypes_visibility - The scope at which the extensions are visible and usable in CloudFormation operations.

Valid values include:

PRIVATE: Extensions that are visible and usable within this account and region. This includes:

Private extensions you have registered in this account and region.

Public extensions that you have activated in this account and region.

PUBLIC: Extensions that are publicly visible and available to be activated within any Amazon account. This includes extensions from Amazon, as well as third-party publishers.

The default is PRIVATE.

$sel:filters:ListTypes', listTypes_filters - Filter criteria to use in determining which extensions to return.

If you specify a filter, CloudFormation ignores any specified Visibility value when returning the list of types.

$sel:type':ListTypes', listTypes_type - The type of extension.

Request Lenses

listTypes_nextToken :: Lens' ListTypes (Maybe Text)

If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

listTypes_maxResults :: Lens' ListTypes (Maybe Natural)

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

listTypes_deprecatedStatus :: Lens' ListTypes (Maybe DeprecatedStatus)

The deprecation status of the extension that you want to get summary information about.

Valid values include:

LIVE: The extension is registered for use in CloudFormation operations.

DEPRECATED: The extension has been deregistered and can no longer be used in CloudFormation operations.

listTypes_provisioningType :: Lens' ListTypes (Maybe ProvisioningType)

For resource types, the provisioning behavior of the resource type. CloudFormation determines the provisioning type during registration, based on the types of handlers in the schema handler package submitted.

Valid values include:

FULLY_MUTABLE: The resource type includes an update handler to process updates to the type during stack update operations.

IMMUTABLE: The resource type does not include an update handler, so the type cannot be updated and must instead be replaced during stack update operations.

NON_PROVISIONABLE: The resource type does not include create, read, and delete handlers, and therefore cannot actually be provisioned.

The default is FULLY_MUTABLE.

listTypes_visibility :: Lens' ListTypes (Maybe Visibility)

The scope at which the extensions are visible and usable in CloudFormation operations.

Valid values include:

PRIVATE: Extensions that are visible and usable within this account and region. This includes:

Private extensions you have registered in this account and region.

Public extensions that you have activated in this account and region.

PUBLIC: Extensions that are publicly visible and available to be activated within any Amazon account. This includes extensions from Amazon, as well as third-party publishers.

The default is PRIVATE.

listTypes_filters :: Lens' ListTypes (Maybe TypeFilters)

Filter criteria to use in determining which extensions to return.

If you specify a filter, CloudFormation ignores any specified Visibility value when returning the list of types.

listTypes_type :: Lens' ListTypes (Maybe RegistryType)

The type of extension.

Destructuring the Response

data ListTypesResponse

See: newListTypesResponse smart constructor.

Constructors

ListTypesResponse'

Fields

nextToken :: Maybe Text

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

typeSummaries :: Maybe [TypeSummary]

A list of TypeSummary structures that contain information about the specified extensions.

httpStatus :: Int

The response's http status code.

newListTypesResponse

Arguments

:: Int
$sel:httpStatus:ListTypesResponse'

-> ListTypesResponse

Create a value of ListTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTypes', listTypesResponse_nextToken - If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

$sel:typeSummaries:ListTypesResponse', listTypesResponse_typeSummaries - A list of TypeSummary structures that contain information about the specified extensions.

$sel:httpStatus:ListTypesResponse', listTypesResponse_httpStatus - The response's http status code.

Response Lenses

listTypesResponse_nextToken :: Lens' ListTypesResponse (Maybe Text)

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call this action again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null.

listTypesResponse_typeSummaries :: Lens' ListTypesResponse (Maybe [TypeSummary])

A list of TypeSummary structures that contain information about the specified extensions.

listTypesResponse_httpStatus :: Lens' ListTypesResponse Int

The response's http status code.