Network.AWS.CloudFormation.PublishType

Description

Publishes the specified extension to the CloudFormation registry as a public extension in this region. Public extensions are available for use by all CloudFormation users. For more information on publishing extensions, see Publishing extensions to make them available for public use in the CloudFormation CLI User Guide.

To publish an extension, you must be registered as a publisher with CloudFormation. For more information, see RegisterPublisher.

Creating a Request

data PublishType

See: newPublishType smart constructor.

Constructors

PublishType'

Fields

typeName :: Maybe Text

The name of the extension.

Conditional: You must specify Arn, or TypeName and Type.

arn :: Maybe Text

The Amazon Resource Number (ARN) of the extension.

Conditional: You must specify Arn, or TypeName and Type.

publicVersionNumber :: Maybe Text

The version number to assign to this version of the extension.

Use the following format, and adhere to semantic versioning when assigning a version number to your extension:

MAJOR.MINOR.PATCH

For more information, see Semantic Versioning 2.0.0.

If you do not specify a version number, CloudFormation increments the version number by one minor version release.

The first time you publish a type, CloudFormation sets the version number to 1.0.0, regardless of the value you specify.

type' :: Maybe ThirdPartyType

The type of the extension.

Conditional: You must specify Arn, or TypeName and Type.

newPublishType :: PublishType

Create a value of PublishType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:PublishType', publishType_typeName - The name of the extension.

Conditional: You must specify Arn, or TypeName and Type.

$sel:arn:PublishType', publishType_arn - The Amazon Resource Number (ARN) of the extension.

Conditional: You must specify Arn, or TypeName and Type.

$sel:publicVersionNumber:PublishType', publishType_publicVersionNumber - The version number to assign to this version of the extension.

Use the following format, and adhere to semantic versioning when assigning a version number to your extension:

MAJOR.MINOR.PATCH

For more information, see Semantic Versioning 2.0.0.

If you do not specify a version number, CloudFormation increments the version number by one minor version release.

The first time you publish a type, CloudFormation sets the version number to 1.0.0, regardless of the value you specify.

$sel:type':PublishType', publishType_type - The type of the extension.

Conditional: You must specify Arn, or TypeName and Type.

Request Lenses

publishType_typeName :: Lens' PublishType (Maybe Text)

The name of the extension.

Conditional: You must specify Arn, or TypeName and Type.

publishType_arn :: Lens' PublishType (Maybe Text)

The Amazon Resource Number (ARN) of the extension.

Conditional: You must specify Arn, or TypeName and Type.

publishType_publicVersionNumber :: Lens' PublishType (Maybe Text)

The version number to assign to this version of the extension.

Use the following format, and adhere to semantic versioning when assigning a version number to your extension:

MAJOR.MINOR.PATCH

For more information, see Semantic Versioning 2.0.0.

If you do not specify a version number, CloudFormation increments the version number by one minor version release.

The first time you publish a type, CloudFormation sets the version number to 1.0.0, regardless of the value you specify.

publishType_type :: Lens' PublishType (Maybe ThirdPartyType)

The type of the extension.

Conditional: You must specify Arn, or TypeName and Type.

Destructuring the Response

data PublishTypeResponse

See: newPublishTypeResponse smart constructor.

Constructors

PublishTypeResponse'

Fields

publicTypeArn :: Maybe Text

The Amazon Resource Number (ARN) assigned to the public extension upon publication.

httpStatus :: Int

The response's http status code.

newPublishTypeResponse

Arguments

:: Int
$sel:httpStatus:PublishTypeResponse'

-> PublishTypeResponse

Create a value of PublishTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:publicTypeArn:PublishTypeResponse', publishTypeResponse_publicTypeArn - The Amazon Resource Number (ARN) assigned to the public extension upon publication.

$sel:httpStatus:PublishTypeResponse', publishTypeResponse_httpStatus - The response's http status code.

Response Lenses

publishTypeResponse_publicTypeArn :: Lens' PublishTypeResponse (Maybe Text)

The Amazon Resource Number (ARN) assigned to the public extension upon publication.

publishTypeResponse_httpStatus :: Lens' PublishTypeResponse Int

The response's http status code.