Network.AWS.CloudFormation.RecordHandlerProgress

Description

Reports progress of a resource handler to CloudFormation.

Reserved for use by the CloudFormation CLI. Do not use this API in your code.

Creating a Request

data RecordHandlerProgress

See: newRecordHandlerProgress smart constructor.

Constructors

RecordHandlerProgress'

Fields

resourceModel :: Maybe Text

Reserved for use by the CloudFormation CLI.

statusMessage :: Maybe Text

Reserved for use by the CloudFormation CLI.

errorCode :: Maybe HandlerErrorCode

Reserved for use by the CloudFormation CLI.

currentOperationStatus :: Maybe OperationStatus

Reserved for use by the CloudFormation CLI.

clientRequestToken :: Maybe Text

Reserved for use by the CloudFormation CLI.

bearerToken :: Text

Reserved for use by the CloudFormation CLI.

operationStatus :: OperationStatus

Reserved for use by the CloudFormation CLI.

newRecordHandlerProgress

Arguments

:: Text
$sel:bearerToken:RecordHandlerProgress'

-> OperationStatus
$sel:operationStatus:RecordHandlerProgress'

-> RecordHandlerProgress

Create a value of RecordHandlerProgress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceModel:RecordHandlerProgress', recordHandlerProgress_resourceModel - Reserved for use by the CloudFormation CLI.

$sel:statusMessage:RecordHandlerProgress', recordHandlerProgress_statusMessage - Reserved for use by the CloudFormation CLI.

$sel:errorCode:RecordHandlerProgress', recordHandlerProgress_errorCode - Reserved for use by the CloudFormation CLI.

$sel:currentOperationStatus:RecordHandlerProgress', recordHandlerProgress_currentOperationStatus - Reserved for use by the CloudFormation CLI.

$sel:clientRequestToken:RecordHandlerProgress', recordHandlerProgress_clientRequestToken - Reserved for use by the CloudFormation CLI.

$sel:bearerToken:RecordHandlerProgress', recordHandlerProgress_bearerToken - Reserved for use by the CloudFormation CLI.

$sel:operationStatus:RecordHandlerProgress', recordHandlerProgress_operationStatus - Reserved for use by the CloudFormation CLI.

Request Lenses

recordHandlerProgress_resourceModel :: Lens' RecordHandlerProgress (Maybe Text)

Reserved for use by the CloudFormation CLI.

recordHandlerProgress_statusMessage :: Lens' RecordHandlerProgress (Maybe Text)

Reserved for use by the CloudFormation CLI.

recordHandlerProgress_errorCode :: Lens' RecordHandlerProgress (Maybe HandlerErrorCode)

Reserved for use by the CloudFormation CLI.

recordHandlerProgress_currentOperationStatus :: Lens' RecordHandlerProgress (Maybe OperationStatus)

Reserved for use by the CloudFormation CLI.

recordHandlerProgress_clientRequestToken :: Lens' RecordHandlerProgress (Maybe Text)

Reserved for use by the CloudFormation CLI.

recordHandlerProgress_bearerToken :: Lens' RecordHandlerProgress Text

Reserved for use by the CloudFormation CLI.

recordHandlerProgress_operationStatus :: Lens' RecordHandlerProgress OperationStatus

Reserved for use by the CloudFormation CLI.

Destructuring the Response

data RecordHandlerProgressResponse

See: newRecordHandlerProgressResponse smart constructor.

Constructors

RecordHandlerProgressResponse'

Fields

httpStatus :: Int

The response's http status code.

newRecordHandlerProgressResponse

Arguments

:: Int
$sel:httpStatus:RecordHandlerProgressResponse'

-> RecordHandlerProgressResponse

Create a value of RecordHandlerProgressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RecordHandlerProgressResponse', recordHandlerProgressResponse_httpStatus - The response's http status code.

Response Lenses

recordHandlerProgressResponse_httpStatus :: Lens' RecordHandlerProgressResponse Int

The response's http status code.