Network.AWS.CloudFormation.RegisterType

Description

Registers an extension with the CloudFormation service. Registering an extension makes it available for use in CloudFormation templates in your Amazon Web Services account, and includes:

Validating the extension schema

Determining which handlers, if any, have been specified for the extension

Making the extension available for use in your account

For more information on how to develop extensions and ready them for registeration, see Creating Resource Providers in the CloudFormation CLI User Guide.

You can have a maximum of 50 resource extension versions registered at a time. This maximum is per account and per region. Use DeregisterType to deregister specific extension versions if necessary.

Once you have initiated a registration request using RegisterType , you can use DescribeTypeRegistration to monitor the progress of the registration request.

Once you have registered a private extension in your account and region, use SetTypeConfiguration to specify configuration properties for the extension. For more information, see Configuring extensions at the account level in the CloudFormation User Guide.

Creating a Request

data RegisterType

See: newRegisterType smart constructor.

Constructors

RegisterType'

Fields

loggingConfig :: Maybe LoggingConfig

Specifies logging configuration information for an extension.

executionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role for CloudFormation to assume when invoking the extension.

For CloudFormation to assume the specified execution role, the role must contain a trust relationship with the CloudFormation service principle (resources.cloudformation.amazonaws.com). For more information on adding trust relationships, see Modifying a role trust policy in the Identity and Access Management User Guide.

If your extension calls Amazon Web Services APIs in any of its handlers, you must create an <https:docs.aws.amazon.comIAMlatestUserGuideid_roles.html IAM execution role> that includes the necessary permissions to call those Amazon Web Services APIs, and provision that execution role in your account. When CloudFormation needs to invoke the resource type handler, CloudFormation assumes this execution role to create a temporary session token, which it then passes to the resource type handler, thereby supplying your resource type with the appropriate credentials.

clientRequestToken :: Maybe Text

A unique identifier that acts as an idempotency key for this registration request. Specifying a client request token prevents CloudFormation from generating more than one version of an extension from the same registeration request, even if the request is submitted multiple times.

type' :: Maybe RegistryType

The kind of extension.

typeName :: Text

The name of the extension being registered.

We recommend that extension names adhere to the following patterns:

For resource types, company_or_organization::service::type.

For modules, company_or_organization::service::type::MODULE.

The following organization namespaces are reserved and cannot be used in your extension names:

Alexa

AMZN

Amazon

AWS

Custom

Dev

schemaHandlerPackage :: Text

A url to the S3 bucket containing the extension project package that contains the neccessary files for the extension you want to register.

For information on generating a schema handler package for the extension you want to register, see submit in the CloudFormation CLI User Guide.

The user registering the extension must be able to access the package in the S3 bucket. That is, the user needs to have GetObject permissions for the schema handler package. For more information, see Actions, Resources, and Condition Keys for Amazon S3 in the Identity and Access Management User Guide.

newRegisterType

Arguments

:: Text
$sel:typeName:RegisterType'

-> Text
$sel:schemaHandlerPackage:RegisterType'

-> RegisterType

Create a value of RegisterType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loggingConfig:RegisterType', registerType_loggingConfig - Specifies logging configuration information for an extension.

$sel:executionRoleArn:RegisterType', registerType_executionRoleArn - The Amazon Resource Name (ARN) of the IAM role for CloudFormation to assume when invoking the extension.

For CloudFormation to assume the specified execution role, the role must contain a trust relationship with the CloudFormation service principle (resources.cloudformation.amazonaws.com). For more information on adding trust relationships, see Modifying a role trust policy in the Identity and Access Management User Guide.

If your extension calls Amazon Web Services APIs in any of its handlers, you must create an <https:docs.aws.amazon.comIAMlatestUserGuideid_roles.html IAM execution role> that includes the necessary permissions to call those Amazon Web Services APIs, and provision that execution role in your account. When CloudFormation needs to invoke the resource type handler, CloudFormation assumes this execution role to create a temporary session token, which it then passes to the resource type handler, thereby supplying your resource type with the appropriate credentials.

$sel:clientRequestToken:RegisterType', registerType_clientRequestToken - A unique identifier that acts as an idempotency key for this registration request. Specifying a client request token prevents CloudFormation from generating more than one version of an extension from the same registeration request, even if the request is submitted multiple times.

$sel:type':RegisterType', registerType_type - The kind of extension.

$sel:typeName:RegisterType', registerType_typeName - The name of the extension being registered.

We recommend that extension names adhere to the following patterns:

For resource types, company_or_organization::service::type.

For modules, company_or_organization::service::type::MODULE.

The following organization namespaces are reserved and cannot be used in your extension names:

Alexa

AMZN

Amazon

AWS

Custom

Dev

$sel:schemaHandlerPackage:RegisterType', registerType_schemaHandlerPackage - A url to the S3 bucket containing the extension project package that contains the neccessary files for the extension you want to register.

For information on generating a schema handler package for the extension you want to register, see submit in the CloudFormation CLI User Guide.

The user registering the extension must be able to access the package in the S3 bucket. That is, the user needs to have GetObject permissions for the schema handler package. For more information, see Actions, Resources, and Condition Keys for Amazon S3 in the Identity and Access Management User Guide.

Request Lenses

registerType_loggingConfig :: Lens' RegisterType (Maybe LoggingConfig)

Specifies logging configuration information for an extension.

registerType_executionRoleArn :: Lens' RegisterType (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role for CloudFormation to assume when invoking the extension.

For CloudFormation to assume the specified execution role, the role must contain a trust relationship with the CloudFormation service principle (resources.cloudformation.amazonaws.com). For more information on adding trust relationships, see Modifying a role trust policy in the Identity and Access Management User Guide.

If your extension calls Amazon Web Services APIs in any of its handlers, you must create an <https:docs.aws.amazon.comIAMlatestUserGuideid_roles.html IAM execution role> that includes the necessary permissions to call those Amazon Web Services APIs, and provision that execution role in your account. When CloudFormation needs to invoke the resource type handler, CloudFormation assumes this execution role to create a temporary session token, which it then passes to the resource type handler, thereby supplying your resource type with the appropriate credentials.

registerType_clientRequestToken :: Lens' RegisterType (Maybe Text)

A unique identifier that acts as an idempotency key for this registration request. Specifying a client request token prevents CloudFormation from generating more than one version of an extension from the same registeration request, even if the request is submitted multiple times.

registerType_type :: Lens' RegisterType (Maybe RegistryType)

The kind of extension.

registerType_typeName :: Lens' RegisterType Text

The name of the extension being registered.

We recommend that extension names adhere to the following patterns:

For resource types, company_or_organization::service::type.

For modules, company_or_organization::service::type::MODULE.

The following organization namespaces are reserved and cannot be used in your extension names:

Alexa

AMZN

Amazon

AWS

Custom

Dev

registerType_schemaHandlerPackage :: Lens' RegisterType Text

A url to the S3 bucket containing the extension project package that contains the neccessary files for the extension you want to register.

For information on generating a schema handler package for the extension you want to register, see submit in the CloudFormation CLI User Guide.

The user registering the extension must be able to access the package in the S3 bucket. That is, the user needs to have GetObject permissions for the schema handler package. For more information, see Actions, Resources, and Condition Keys for Amazon S3 in the Identity and Access Management User Guide.

Destructuring the Response

data RegisterTypeResponse

See: newRegisterTypeResponse smart constructor.

Constructors

RegisterTypeResponse'

Fields

registrationToken :: Maybe Text

The identifier for this registration request.

Use this registration token when calling DescribeTypeRegistration , which returns information about the status and IDs of the extension registration.

httpStatus :: Int

The response's http status code.

newRegisterTypeResponse

Arguments

:: Int
$sel:httpStatus:RegisterTypeResponse'

-> RegisterTypeResponse

Create a value of RegisterTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registrationToken:RegisterTypeResponse', registerTypeResponse_registrationToken - The identifier for this registration request.

Use this registration token when calling DescribeTypeRegistration , which returns information about the status and IDs of the extension registration.

$sel:httpStatus:RegisterTypeResponse', registerTypeResponse_httpStatus - The response's http status code.

Response Lenses

registerTypeResponse_registrationToken :: Lens' RegisterTypeResponse (Maybe Text)

The identifier for this registration request.

Use this registration token when calling DescribeTypeRegistration , which returns information about the status and IDs of the extension registration.

registerTypeResponse_httpStatus :: Lens' RegisterTypeResponse Int

The response's http status code.