Network.AWS.CloudFormation.SetTypeConfiguration

Description

Specifies the configuration data for a registered CloudFormation extension, in the given account and region.

To view the current configuration data for an extension, refer to the ConfigurationSchema element of DescribeType. For more information, see Configuring extensions at the account level in the CloudFormation User Guide.

It is strongly recommended that you use dynamic references to restrict sensitive configuration definitions, such as third-party credentials. For more details on dynamic references, see Using dynamic references to specify template values in the CloudFormation User Guide.

Creating a Request

data SetTypeConfiguration

See: newSetTypeConfiguration smart constructor.

Constructors

SetTypeConfiguration'

Fields

typeName :: Maybe Text

The name of the extension.

Conditional: You must specify ConfigurationArn, or Type and TypeName.

typeArn :: Maybe Text

The Amazon Resource Name (ARN) for the extension, in this account and region.

For public extensions, this will be the ARN assigned when you activate the type in this account and region. For private extensions, this will be the ARN assigned when you register the type in this account and region.

Do not include the extension versions suffix at the end of the ARN. You can set the configuration for an extension, but not for a specific extension version.

type' :: Maybe ThirdPartyType

The type of extension.

Conditional: You must specify ConfigurationArn, or Type and TypeName.

configurationAlias :: Maybe Text

An alias by which to refer to this extension configuration data.

Conditional: Specifying a configuration alias is required when setting a configuration for a resource type extension.

configuration :: Text

The configuration data for the extension, in this account and region.

The configuration data must be formatted as JSON, and validate against the schema returned in the ConfigurationSchema response element of API_DescribeType. For more information, see Defining account-level configuration data for an extension in the CloudFormation CLI User Guide.

newSetTypeConfiguration

Arguments

:: Text
$sel:configuration:SetTypeConfiguration'

-> SetTypeConfiguration

Create a value of SetTypeConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:SetTypeConfiguration', setTypeConfiguration_typeName - The name of the extension.

Conditional: You must specify ConfigurationArn, or Type and TypeName.

$sel:typeArn:SetTypeConfiguration', setTypeConfiguration_typeArn - The Amazon Resource Name (ARN) for the extension, in this account and region.

For public extensions, this will be the ARN assigned when you activate the type in this account and region. For private extensions, this will be the ARN assigned when you register the type in this account and region.

Do not include the extension versions suffix at the end of the ARN. You can set the configuration for an extension, but not for a specific extension version.

$sel:type':SetTypeConfiguration', setTypeConfiguration_type - The type of extension.

Conditional: You must specify ConfigurationArn, or Type and TypeName.

$sel:configurationAlias:SetTypeConfiguration', setTypeConfiguration_configurationAlias - An alias by which to refer to this extension configuration data.

Conditional: Specifying a configuration alias is required when setting a configuration for a resource type extension.

$sel:configuration:SetTypeConfiguration', setTypeConfiguration_configuration - The configuration data for the extension, in this account and region.

The configuration data must be formatted as JSON, and validate against the schema returned in the ConfigurationSchema response element of API_DescribeType. For more information, see Defining account-level configuration data for an extension in the CloudFormation CLI User Guide.

Request Lenses

setTypeConfiguration_typeName :: Lens' SetTypeConfiguration (Maybe Text)

The name of the extension.

Conditional: You must specify ConfigurationArn, or Type and TypeName.

setTypeConfiguration_typeArn :: Lens' SetTypeConfiguration (Maybe Text)

The Amazon Resource Name (ARN) for the extension, in this account and region.

For public extensions, this will be the ARN assigned when you activate the type in this account and region. For private extensions, this will be the ARN assigned when you register the type in this account and region.

Do not include the extension versions suffix at the end of the ARN. You can set the configuration for an extension, but not for a specific extension version.

setTypeConfiguration_type :: Lens' SetTypeConfiguration (Maybe ThirdPartyType)

The type of extension.

Conditional: You must specify ConfigurationArn, or Type and TypeName.

setTypeConfiguration_configurationAlias :: Lens' SetTypeConfiguration (Maybe Text)

An alias by which to refer to this extension configuration data.

Conditional: Specifying a configuration alias is required when setting a configuration for a resource type extension.

setTypeConfiguration_configuration :: Lens' SetTypeConfiguration Text

The configuration data for the extension, in this account and region.

The configuration data must be formatted as JSON, and validate against the schema returned in the ConfigurationSchema response element of API_DescribeType. For more information, see Defining account-level configuration data for an extension in the CloudFormation CLI User Guide.

Destructuring the Response

data SetTypeConfigurationResponse

See: newSetTypeConfigurationResponse smart constructor.

Constructors

SetTypeConfigurationResponse'

Fields

configurationArn :: Maybe Text

The Amazon Resource Name (ARN) for the configuration data, in this account and region.

Conditional: You must specify ConfigurationArn, or Type and TypeName.

httpStatus :: Int

The response's http status code.

newSetTypeConfigurationResponse

Arguments

:: Int
$sel:httpStatus:SetTypeConfigurationResponse'

-> SetTypeConfigurationResponse

Create a value of SetTypeConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationArn:SetTypeConfigurationResponse', setTypeConfigurationResponse_configurationArn - The Amazon Resource Name (ARN) for the configuration data, in this account and region.

Conditional: You must specify ConfigurationArn, or Type and TypeName.

$sel:httpStatus:SetTypeConfigurationResponse', setTypeConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

setTypeConfigurationResponse_configurationArn :: Lens' SetTypeConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) for the configuration data, in this account and region.

Conditional: You must specify ConfigurationArn, or Type and TypeName.

setTypeConfigurationResponse_httpStatus :: Lens' SetTypeConfigurationResponse Int

The response's http status code.