Network.AWS.CloudFormation.SignalResource

Description

Sends a signal to the specified resource with a success or failure status. You can use the SignalResource API in conjunction with a creation policy or update policy. CloudFormation doesn't proceed with a stack creation or update until resources receive the required number of signals or the timeout period is exceeded. The SignalResource API is useful in cases where you want to send signals from anywhere other than an Amazon EC2 instance.

Creating a Request

data SignalResource

The input for the SignalResource action.

See: newSignalResource smart constructor.

Constructors

SignalResource'

Fields

stackName :: Text

The stack name or unique stack ID that includes the resource that you want to signal.

logicalResourceId :: Text

The logical ID of the resource that you want to signal. The logical ID is the name of the resource that given in the template.

uniqueId :: Text

A unique ID of the signal. When you signal Amazon EC2 instances or Auto Scaling groups, specify the instance ID that you are signaling as the unique ID. If you send multiple signals to a single resource (such as signaling a wait condition), each signal requires a different unique ID.

status :: ResourceSignalStatus

The status of the signal, which is either success or failure. A failure signal causes CloudFormation to immediately fail the stack creation or update.

newSignalResource

Arguments

:: Text
$sel:stackName:SignalResource'

-> Text
$sel:logicalResourceId:SignalResource'

-> Text
$sel:uniqueId:SignalResource'

-> ResourceSignalStatus
$sel:status:SignalResource'

-> SignalResource

Create a value of SignalResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stackName:SignalResource', signalResource_stackName - The stack name or unique stack ID that includes the resource that you want to signal.

$sel:logicalResourceId:SignalResource', signalResource_logicalResourceId - The logical ID of the resource that you want to signal. The logical ID is the name of the resource that given in the template.

$sel:uniqueId:SignalResource', signalResource_uniqueId - A unique ID of the signal. When you signal Amazon EC2 instances or Auto Scaling groups, specify the instance ID that you are signaling as the unique ID. If you send multiple signals to a single resource (such as signaling a wait condition), each signal requires a different unique ID.

$sel:status:SignalResource', signalResource_status - The status of the signal, which is either success or failure. A failure signal causes CloudFormation to immediately fail the stack creation or update.

Request Lenses

signalResource_stackName :: Lens' SignalResource Text

The stack name or unique stack ID that includes the resource that you want to signal.

signalResource_logicalResourceId :: Lens' SignalResource Text

The logical ID of the resource that you want to signal. The logical ID is the name of the resource that given in the template.

signalResource_uniqueId :: Lens' SignalResource Text

A unique ID of the signal. When you signal Amazon EC2 instances or Auto Scaling groups, specify the instance ID that you are signaling as the unique ID. If you send multiple signals to a single resource (such as signaling a wait condition), each signal requires a different unique ID.

signalResource_status :: Lens' SignalResource ResourceSignalStatus

The status of the signal, which is either success or failure. A failure signal causes CloudFormation to immediately fail the stack creation or update.

Destructuring the Response

data SignalResourceResponse

See: newSignalResourceResponse smart constructor.

Constructors

SignalResourceResponse'

newSignalResourceResponse :: SignalResourceResponse

Create a value of SignalResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.