Network.AWS.CloudFormation.StopStackSetOperation

Description

Stops an in-progress operation on a stack set and its associated stack instances.

Creating a Request

data StopStackSetOperation

See: newStopStackSetOperation smart constructor.

Constructors

StopStackSetOperation'

Fields

callAs :: Maybe CallAs

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

stackSetName :: Text

The name or unique ID of the stack set that you want to stop the operation for.

operationId :: Text

The ID of the stack operation.

newStopStackSetOperation

Arguments

:: Text
$sel:stackSetName:StopStackSetOperation'

-> Text
$sel:operationId:StopStackSetOperation'

-> StopStackSetOperation

Create a value of StopStackSetOperation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:callAs:StopStackSetOperation', stopStackSetOperation_callAs - [Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

$sel:stackSetName:StopStackSetOperation', stopStackSetOperation_stackSetName - The name or unique ID of the stack set that you want to stop the operation for.

$sel:operationId:StopStackSetOperation', stopStackSetOperation_operationId - The ID of the stack operation.

Request Lenses

stopStackSetOperation_callAs :: Lens' StopStackSetOperation (Maybe CallAs)

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

stopStackSetOperation_stackSetName :: Lens' StopStackSetOperation Text

The name or unique ID of the stack set that you want to stop the operation for.

stopStackSetOperation_operationId :: Lens' StopStackSetOperation Text

The ID of the stack operation.

Destructuring the Response

data StopStackSetOperationResponse

See: newStopStackSetOperationResponse smart constructor.

Constructors

StopStackSetOperationResponse'

Fields

httpStatus :: Int

The response's http status code.

newStopStackSetOperationResponse

Arguments

:: Int
$sel:httpStatus:StopStackSetOperationResponse'

-> StopStackSetOperationResponse

Create a value of StopStackSetOperationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StopStackSetOperationResponse', stopStackSetOperationResponse_httpStatus - The response's http status code.

Response Lenses

stopStackSetOperationResponse_httpStatus :: Lens' StopStackSetOperationResponse Int

The response's http status code.