Network.AWS.CloudFormation.TestType

Description

Tests a registered extension to make sure it meets all necessary requirements for being published in the CloudFormation registry.

For resource types, this includes passing all contracts tests defined for the type.

For modules, this includes determining if the module's model meets all necessary requirements.

For more information, see Testing your public extension prior to publishing in the CloudFormation CLI User Guide.

If you do not specify a version, CloudFormation uses the default version of the extension in your account and region for testing.

To perform testing, CloudFormation assumes the execution role specified when the type was registered. For more information, see RegisterType.

Once you've initiated testing on an extension using TestType, you can use DescribeType to monitor the current test status and test status description for the extension.

An extension must have a test status of PASSED before it can be published. For more information, see Publishing extensions to make them available for public use in the CloudFormation CLI User Guide.

Creating a Request

data TestType

See: newTestType smart constructor.

Constructors

TestType'

Fields

typeName :: Maybe Text

The name of the extension to test.

Conditional: You must specify Arn, or TypeName and Type.

logDeliveryBucket :: Maybe Text

The S3 bucket to which CloudFormation delivers the contract test execution logs.

CloudFormation delivers the logs by the time contract testing has completed and the extension has been assigned a test type status of PASSED or FAILED.

The user calling TestType must be able to access items in the specified S3 bucket. Specifically, the user needs the following permissions:

GetObject

PutObject

For more information, see Actions, Resources, and Condition Keys for Amazon S3 in the Amazon Web Services Identity and Access Management User Guide.

arn :: Maybe Text

The Amazon Resource Number (ARN) of the extension.

Conditional: You must specify Arn, or TypeName and Type.

versionId :: Maybe Text

The version of the extension to test.

You can specify the version id with either Arn, or with TypeName and Type.

If you do not specify a version, CloudFormation uses the default version of the extension in this account and region for testing.

type' :: Maybe ThirdPartyType

The type of the extension to test.

Conditional: You must specify Arn, or TypeName and Type.

newTestType :: TestType

Create a value of TestType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:TestType', testType_typeName - The name of the extension to test.

Conditional: You must specify Arn, or TypeName and Type.

$sel:logDeliveryBucket:TestType', testType_logDeliveryBucket - The S3 bucket to which CloudFormation delivers the contract test execution logs.

CloudFormation delivers the logs by the time contract testing has completed and the extension has been assigned a test type status of PASSED or FAILED.

The user calling TestType must be able to access items in the specified S3 bucket. Specifically, the user needs the following permissions:

GetObject

PutObject

For more information, see Actions, Resources, and Condition Keys for Amazon S3 in the Amazon Web Services Identity and Access Management User Guide.

$sel:arn:TestType', testType_arn - The Amazon Resource Number (ARN) of the extension.

Conditional: You must specify Arn, or TypeName and Type.

$sel:versionId:TestType', testType_versionId - The version of the extension to test.

You can specify the version id with either Arn, or with TypeName and Type.

If you do not specify a version, CloudFormation uses the default version of the extension in this account and region for testing.

$sel:type':TestType', testType_type - The type of the extension to test.

Conditional: You must specify Arn, or TypeName and Type.

Request Lenses

testType_typeName :: Lens' TestType (Maybe Text)

The name of the extension to test.

Conditional: You must specify Arn, or TypeName and Type.

testType_logDeliveryBucket :: Lens' TestType (Maybe Text)

The S3 bucket to which CloudFormation delivers the contract test execution logs.

CloudFormation delivers the logs by the time contract testing has completed and the extension has been assigned a test type status of PASSED or FAILED.

The user calling TestType must be able to access items in the specified S3 bucket. Specifically, the user needs the following permissions:

GetObject

PutObject

For more information, see Actions, Resources, and Condition Keys for Amazon S3 in the Amazon Web Services Identity and Access Management User Guide.

testType_arn :: Lens' TestType (Maybe Text)

The Amazon Resource Number (ARN) of the extension.

Conditional: You must specify Arn, or TypeName and Type.

testType_versionId :: Lens' TestType (Maybe Text)

The version of the extension to test.

You can specify the version id with either Arn, or with TypeName and Type.

If you do not specify a version, CloudFormation uses the default version of the extension in this account and region for testing.

testType_type :: Lens' TestType (Maybe ThirdPartyType)

The type of the extension to test.

Conditional: You must specify Arn, or TypeName and Type.

Destructuring the Response

data TestTypeResponse

See: newTestTypeResponse smart constructor.

Constructors

TestTypeResponse'

Fields

typeVersionArn :: Maybe Text

The Amazon Resource Number (ARN) of the extension.

httpStatus :: Int

The response's http status code.

newTestTypeResponse

Arguments

:: Int
$sel:httpStatus:TestTypeResponse'

-> TestTypeResponse

Create a value of TestTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeVersionArn:TestTypeResponse', testTypeResponse_typeVersionArn - The Amazon Resource Number (ARN) of the extension.

$sel:httpStatus:TestTypeResponse', testTypeResponse_httpStatus - The response's http status code.

Response Lenses

testTypeResponse_typeVersionArn :: Lens' TestTypeResponse (Maybe Text)

The Amazon Resource Number (ARN) of the extension.

testTypeResponse_httpStatus :: Lens' TestTypeResponse Int

The response's http status code.