Network.AWS.CloudFormation.Types.BatchDescribeTypeConfigurationsError

Description

Documentation

data BatchDescribeTypeConfigurationsError

Detailed information concerning an error generated during the setting of configuration data for a CloudFormation extension.

See: newBatchDescribeTypeConfigurationsError smart constructor.

Constructors

BatchDescribeTypeConfigurationsError'

Fields

typeConfigurationIdentifier :: Maybe TypeConfigurationIdentifier

errorMessage :: Maybe Text

The error message.

errorCode :: Maybe Text

The error code.

newBatchDescribeTypeConfigurationsError :: BatchDescribeTypeConfigurationsError

Create a value of BatchDescribeTypeConfigurationsError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeConfigurationIdentifier:BatchDescribeTypeConfigurationsError', batchDescribeTypeConfigurationsError_typeConfigurationIdentifier - Undocumented member.

$sel:errorMessage:BatchDescribeTypeConfigurationsError', batchDescribeTypeConfigurationsError_errorMessage - The error message.

$sel:errorCode:BatchDescribeTypeConfigurationsError', batchDescribeTypeConfigurationsError_errorCode - The error code.

batchDescribeTypeConfigurationsError_typeConfigurationIdentifier :: Lens' BatchDescribeTypeConfigurationsError (Maybe TypeConfigurationIdentifier)

Undocumented member.

batchDescribeTypeConfigurationsError_errorMessage :: Lens' BatchDescribeTypeConfigurationsError (Maybe Text)

The error message.

batchDescribeTypeConfigurationsError_errorCode :: Lens' BatchDescribeTypeConfigurationsError (Maybe Text)

The error code.