Network.AWS.CloudFormation.Types.ChangeSetSummary

Description

Documentation

data ChangeSetSummary

The ChangeSetSummary structure describes a change set, its status, and the stack with which it's associated.

See: newChangeSetSummary smart constructor.

Constructors

ChangeSetSummary'

Fields

creationTime :: Maybe ISO8601

The start time when the change set was created, in UTC.

status :: Maybe ChangeSetStatus

The state of the change set, such as CREATE_IN_PROGRESS, CREATE_COMPLETE, or FAILED.

parentChangeSetId :: Maybe Text

The parent change set ID.

changeSetName :: Maybe Text

The name of the change set.

executionStatus :: Maybe ExecutionStatus

If the change set execution status is AVAILABLE, you can execute the change set. If you can’t execute the change set, the status indicates why. For example, a change set might be in an UNAVAILABLE state because CloudFormation is still creating it or in an OBSOLETE state because the stack was already updated.

changeSetId :: Maybe Text

The ID of the change set.

includeNestedStacks :: Maybe Bool

Specifies the current setting of IncludeNestedStacks for the change set.

rootChangeSetId :: Maybe Text

The root change set ID.

statusReason :: Maybe Text

A description of the change set's status. For example, if your change set is in the FAILED state, CloudFormation shows the error message.

stackId :: Maybe Text

The ID of the stack with which the change set is associated.

description :: Maybe Text

Descriptive information about the change set.

stackName :: Maybe Text

The name of the stack with which the change set is associated.

newChangeSetSummary :: ChangeSetSummary

Create a value of ChangeSetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ChangeSetSummary', changeSetSummary_creationTime - The start time when the change set was created, in UTC.

$sel:status:ChangeSetSummary', changeSetSummary_status - The state of the change set, such as CREATE_IN_PROGRESS, CREATE_COMPLETE, or FAILED.

$sel:parentChangeSetId:ChangeSetSummary', changeSetSummary_parentChangeSetId - The parent change set ID.

$sel:changeSetName:ChangeSetSummary', changeSetSummary_changeSetName - The name of the change set.

$sel:executionStatus:ChangeSetSummary', changeSetSummary_executionStatus - If the change set execution status is AVAILABLE, you can execute the change set. If you can’t execute the change set, the status indicates why. For example, a change set might be in an UNAVAILABLE state because CloudFormation is still creating it or in an OBSOLETE state because the stack was already updated.

$sel:changeSetId:ChangeSetSummary', changeSetSummary_changeSetId - The ID of the change set.

$sel:includeNestedStacks:ChangeSetSummary', changeSetSummary_includeNestedStacks - Specifies the current setting of IncludeNestedStacks for the change set.

$sel:rootChangeSetId:ChangeSetSummary', changeSetSummary_rootChangeSetId - The root change set ID.

$sel:statusReason:ChangeSetSummary', changeSetSummary_statusReason - A description of the change set's status. For example, if your change set is in the FAILED state, CloudFormation shows the error message.

$sel:stackId:ChangeSetSummary', changeSetSummary_stackId - The ID of the stack with which the change set is associated.

$sel:description:ChangeSetSummary', changeSetSummary_description - Descriptive information about the change set.

$sel:stackName:ChangeSetSummary', changeSetSummary_stackName - The name of the stack with which the change set is associated.

changeSetSummary_creationTime :: Lens' ChangeSetSummary (Maybe UTCTime)

The start time when the change set was created, in UTC.

changeSetSummary_status :: Lens' ChangeSetSummary (Maybe ChangeSetStatus)

The state of the change set, such as CREATE_IN_PROGRESS, CREATE_COMPLETE, or FAILED.

changeSetSummary_parentChangeSetId :: Lens' ChangeSetSummary (Maybe Text)

The parent change set ID.

changeSetSummary_changeSetName :: Lens' ChangeSetSummary (Maybe Text)

The name of the change set.

changeSetSummary_executionStatus :: Lens' ChangeSetSummary (Maybe ExecutionStatus)

If the change set execution status is AVAILABLE, you can execute the change set. If you can’t execute the change set, the status indicates why. For example, a change set might be in an UNAVAILABLE state because CloudFormation is still creating it or in an OBSOLETE state because the stack was already updated.

changeSetSummary_changeSetId :: Lens' ChangeSetSummary (Maybe Text)

The ID of the change set.

changeSetSummary_includeNestedStacks :: Lens' ChangeSetSummary (Maybe Bool)

Specifies the current setting of IncludeNestedStacks for the change set.

changeSetSummary_rootChangeSetId :: Lens' ChangeSetSummary (Maybe Text)

The root change set ID.

changeSetSummary_statusReason :: Lens' ChangeSetSummary (Maybe Text)

A description of the change set's status. For example, if your change set is in the FAILED state, CloudFormation shows the error message.

changeSetSummary_stackId :: Lens' ChangeSetSummary (Maybe Text)

The ID of the stack with which the change set is associated.

changeSetSummary_description :: Lens' ChangeSetSummary (Maybe Text)

Descriptive information about the change set.

changeSetSummary_stackName :: Lens' ChangeSetSummary (Maybe Text)

The name of the stack with which the change set is associated.