Network.AWS.CloudFormation.Types.LoggingConfig

Description

Documentation

data LoggingConfig

Contains logging configuration information for an extension.

See: newLoggingConfig smart constructor.

Constructors

LoggingConfig'

Fields

logRoleArn :: Text

The ARN of the role that CloudFormation should assume when sending log entries to CloudWatch logs.

logGroupName :: Text

The Amazon CloudWatch log group to which CloudFormation sends error logging information when invoking the extension's handlers.

newLoggingConfig

Arguments

:: Text
$sel:logRoleArn:LoggingConfig'

-> Text
$sel:logGroupName:LoggingConfig'

-> LoggingConfig

Create a value of LoggingConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logRoleArn:LoggingConfig', loggingConfig_logRoleArn - The ARN of the role that CloudFormation should assume when sending log entries to CloudWatch logs.

$sel:logGroupName:LoggingConfig', loggingConfig_logGroupName - The Amazon CloudWatch log group to which CloudFormation sends error logging information when invoking the extension's handlers.

loggingConfig_logRoleArn :: Lens' LoggingConfig Text

The ARN of the role that CloudFormation should assume when sending log entries to CloudWatch logs.

loggingConfig_logGroupName :: Lens' LoggingConfig Text

The Amazon CloudWatch log group to which CloudFormation sends error logging information when invoking the extension's handlers.