Network.AWS.CloudFormation.Types.ParameterDeclaration

Description

Documentation

data ParameterDeclaration

The ParameterDeclaration data type.

See: newParameterDeclaration smart constructor.

Constructors

ParameterDeclaration'

Fields

parameterConstraints :: Maybe ParameterConstraints

The criteria that CloudFormation uses to validate parameter values.

parameterType :: Maybe Text

The type of parameter.

description :: Maybe Text

The description that is associate with the parameter.

parameterKey :: Maybe Text

The name that is associated with the parameter.

defaultValue :: Maybe Text

The default value of the parameter.

noEcho :: Maybe Bool

Flag that indicates whether the parameter value is shown as plain text in logs and in the Amazon Web Services Management Console.

newParameterDeclaration :: ParameterDeclaration

Create a value of ParameterDeclaration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterConstraints:ParameterDeclaration', parameterDeclaration_parameterConstraints - The criteria that CloudFormation uses to validate parameter values.

$sel:parameterType:ParameterDeclaration', parameterDeclaration_parameterType - The type of parameter.

$sel:description:ParameterDeclaration', parameterDeclaration_description - The description that is associate with the parameter.

$sel:parameterKey:ParameterDeclaration', parameterDeclaration_parameterKey - The name that is associated with the parameter.

$sel:defaultValue:ParameterDeclaration', parameterDeclaration_defaultValue - The default value of the parameter.

$sel:noEcho:ParameterDeclaration', parameterDeclaration_noEcho - Flag that indicates whether the parameter value is shown as plain text in logs and in the Amazon Web Services Management Console.

parameterDeclaration_parameterConstraints :: Lens' ParameterDeclaration (Maybe ParameterConstraints)

The criteria that CloudFormation uses to validate parameter values.

parameterDeclaration_parameterType :: Lens' ParameterDeclaration (Maybe Text)

The type of parameter.

parameterDeclaration_description :: Lens' ParameterDeclaration (Maybe Text)

The description that is associate with the parameter.

parameterDeclaration_parameterKey :: Lens' ParameterDeclaration (Maybe Text)

The name that is associated with the parameter.

parameterDeclaration_defaultValue :: Lens' ParameterDeclaration (Maybe Text)

The default value of the parameter.

parameterDeclaration_noEcho :: Lens' ParameterDeclaration (Maybe Bool)

Flag that indicates whether the parameter value is shown as plain text in logs and in the Amazon Web Services Management Console.