Network.AWS.CloudFormation.Types.PropertyDifference

Description

Documentation

data PropertyDifference

Information about a resource property whose actual value differs from its expected value, as defined in the stack template and any values specified as template parameters. These will be present only for resources whose StackResourceDriftStatus is MODIFIED. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

See: newPropertyDifference smart constructor.

Constructors

PropertyDifference'

Fields

propertyPath :: Text

The fully-qualified path to the resource property.

expectedValue :: Text

The expected property value of the resource property, as defined in the stack template and any values specified as template parameters.

actualValue :: Text

The actual property value of the resource property.

differenceType :: DifferenceType

The type of property difference.

ADD: A value has been added to a resource property that is an array or list data type.

REMOVE: The property has been removed from the current resource configuration.

NOT_EQUAL: The current property value differs from its expected value (as defined in the stack template and any values specified as template parameters).

newPropertyDifference

Arguments

:: Text
$sel:propertyPath:PropertyDifference'

-> Text
$sel:expectedValue:PropertyDifference'

-> Text
$sel:actualValue:PropertyDifference'

-> DifferenceType
$sel:differenceType:PropertyDifference'

-> PropertyDifference

Create a value of PropertyDifference with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:propertyPath:PropertyDifference', propertyDifference_propertyPath - The fully-qualified path to the resource property.

$sel:expectedValue:PropertyDifference', propertyDifference_expectedValue - The expected property value of the resource property, as defined in the stack template and any values specified as template parameters.

$sel:actualValue:PropertyDifference', propertyDifference_actualValue - The actual property value of the resource property.

$sel:differenceType:PropertyDifference', propertyDifference_differenceType - The type of property difference.

ADD: A value has been added to a resource property that is an array or list data type.

REMOVE: The property has been removed from the current resource configuration.

NOT_EQUAL: The current property value differs from its expected value (as defined in the stack template and any values specified as template parameters).

propertyDifference_propertyPath :: Lens' PropertyDifference Text

The fully-qualified path to the resource property.

propertyDifference_expectedValue :: Lens' PropertyDifference Text

The expected property value of the resource property, as defined in the stack template and any values specified as template parameters.

propertyDifference_actualValue :: Lens' PropertyDifference Text

The actual property value of the resource property.

propertyDifference_differenceType :: Lens' PropertyDifference DifferenceType

The type of property difference.

ADD: A value has been added to a resource property that is an array or list data type.

REMOVE: The property has been removed from the current resource configuration.

NOT_EQUAL: The current property value differs from its expected value (as defined in the stack template and any values specified as template parameters).