Network.AWS.CloudFormation.Types.RequiredActivatedType

Description

Documentation

data RequiredActivatedType

For extensions that are modules, a public third-party extension that must be activated in your account in order for the module itself to be activated.

For more information, see Activating public modules for use in your account in the CloudFormation User Guide.

See: newRequiredActivatedType smart constructor.

Constructors

RequiredActivatedType'

Fields

originalTypeName :: Maybe Text

The type name of the public extension.

If you specified a TypeNameAlias when enabling the extension in this account and region, CloudFormation treats that alias as the extension's type name within the account and region, not the type name of the public extension. For more information, see Specifying aliases to refer to extensions in the CloudFormation User Guide.

typeNameAlias :: Maybe Text

An alias assigned to the public extension, in this account and region. If you specify an alias for the extension, CloudFormation treats the alias as the extension type name within this account and region. You must use the alias to refer to the extension in your templates, API calls, and CloudFormation console.

supportedMajorVersions :: Maybe [Natural]

A list of the major versions of the extension type that the macro supports.

publisherId :: Maybe Text

The publisher ID of the extension publisher.

newRequiredActivatedType :: RequiredActivatedType

Create a value of RequiredActivatedType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:originalTypeName:RequiredActivatedType', requiredActivatedType_originalTypeName - The type name of the public extension.

If you specified a TypeNameAlias when enabling the extension in this account and region, CloudFormation treats that alias as the extension's type name within the account and region, not the type name of the public extension. For more information, see Specifying aliases to refer to extensions in the CloudFormation User Guide.

$sel:typeNameAlias:RequiredActivatedType', requiredActivatedType_typeNameAlias - An alias assigned to the public extension, in this account and region. If you specify an alias for the extension, CloudFormation treats the alias as the extension type name within this account and region. You must use the alias to refer to the extension in your templates, API calls, and CloudFormation console.

$sel:supportedMajorVersions:RequiredActivatedType', requiredActivatedType_supportedMajorVersions - A list of the major versions of the extension type that the macro supports.

$sel:publisherId:RequiredActivatedType', requiredActivatedType_publisherId - The publisher ID of the extension publisher.

requiredActivatedType_originalTypeName :: Lens' RequiredActivatedType (Maybe Text)

The type name of the public extension.

If you specified a TypeNameAlias when enabling the extension in this account and region, CloudFormation treats that alias as the extension's type name within the account and region, not the type name of the public extension. For more information, see Specifying aliases to refer to extensions in the CloudFormation User Guide.

requiredActivatedType_typeNameAlias :: Lens' RequiredActivatedType (Maybe Text)

An alias assigned to the public extension, in this account and region. If you specify an alias for the extension, CloudFormation treats the alias as the extension type name within this account and region. You must use the alias to refer to the extension in your templates, API calls, and CloudFormation console.

requiredActivatedType_supportedMajorVersions :: Lens' RequiredActivatedType (Maybe [Natural])

A list of the major versions of the extension type that the macro supports.

requiredActivatedType_publisherId :: Lens' RequiredActivatedType (Maybe Text)

The publisher ID of the extension publisher.