Network.AWS.CloudFormation.Types.ResourceChange

Description

Documentation

data ResourceChange

The ResourceChange structure describes the resource and the action that CloudFormation will perform on it if you execute this change set.

See: newResourceChange smart constructor.

Constructors

ResourceChange'

Fields

resourceType :: Maybe Text

The type of CloudFormation resource, such as AWS::S3::Bucket.

physicalResourceId :: Maybe Text

The resource's physical ID (resource name). Resources that you are adding don't have physical IDs because they haven't been created.

details :: Maybe [ResourceChangeDetail]

For the Modify action, a list of ResourceChangeDetail structures that describes the changes that CloudFormation will make to the resource.

scope :: Maybe [ResourceAttribute]

For the Modify action, indicates which resource attribute is triggering this update, such as a change in the resource attribute's Metadata, Properties, or Tags.

moduleInfo :: Maybe ModuleInfo

Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

logicalResourceId :: Maybe Text

The resource's logical ID, which is defined in the stack's template.

changeSetId :: Maybe Text

The change set ID of the nested change set.

action :: Maybe ChangeAction

The action that CloudFormation takes on the resource, such as Add (adds a new resource), Modify (changes a resource), Remove (deletes a resource), Import (imports a resource), or Dynamic (exact action for the resource cannot be determined).

replacement :: Maybe Replacement

For the Modify action, indicates whether CloudFormation will replace the resource by creating a new one and deleting the old one. This value depends on the value of the RequiresRecreation property in the ResourceTargetDefinition structure. For example, if the RequiresRecreation field is Always and the Evaluation field is Static, Replacement is True. If the RequiresRecreation field is Always and the Evaluation field is Dynamic, Replacement is Conditionally.

If you have multiple changes with different RequiresRecreation values, the Replacement value depends on the change with the most impact. A RequiresRecreation value of Always has the most impact, followed by Conditionally, and then Never.

newResourceChange :: ResourceChange

Create a value of ResourceChange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ResourceChange', resourceChange_resourceType - The type of CloudFormation resource, such as AWS::S3::Bucket.

$sel:physicalResourceId:ResourceChange', resourceChange_physicalResourceId - The resource's physical ID (resource name). Resources that you are adding don't have physical IDs because they haven't been created.

$sel:details:ResourceChange', resourceChange_details - For the Modify action, a list of ResourceChangeDetail structures that describes the changes that CloudFormation will make to the resource.

$sel:scope:ResourceChange', resourceChange_scope - For the Modify action, indicates which resource attribute is triggering this update, such as a change in the resource attribute's Metadata, Properties, or Tags.

$sel:moduleInfo:ResourceChange', resourceChange_moduleInfo - Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

$sel:logicalResourceId:ResourceChange', resourceChange_logicalResourceId - The resource's logical ID, which is defined in the stack's template.

$sel:changeSetId:ResourceChange', resourceChange_changeSetId - The change set ID of the nested change set.

$sel:action:ResourceChange', resourceChange_action - The action that CloudFormation takes on the resource, such as Add (adds a new resource), Modify (changes a resource), Remove (deletes a resource), Import (imports a resource), or Dynamic (exact action for the resource cannot be determined).

$sel:replacement:ResourceChange', resourceChange_replacement - For the Modify action, indicates whether CloudFormation will replace the resource by creating a new one and deleting the old one. This value depends on the value of the RequiresRecreation property in the ResourceTargetDefinition structure. For example, if the RequiresRecreation field is Always and the Evaluation field is Static, Replacement is True. If the RequiresRecreation field is Always and the Evaluation field is Dynamic, Replacement is Conditionally.

If you have multiple changes with different RequiresRecreation values, the Replacement value depends on the change with the most impact. A RequiresRecreation value of Always has the most impact, followed by Conditionally, and then Never.

resourceChange_resourceType :: Lens' ResourceChange (Maybe Text)

The type of CloudFormation resource, such as AWS::S3::Bucket.

resourceChange_physicalResourceId :: Lens' ResourceChange (Maybe Text)

The resource's physical ID (resource name). Resources that you are adding don't have physical IDs because they haven't been created.

resourceChange_details :: Lens' ResourceChange (Maybe [ResourceChangeDetail])

For the Modify action, a list of ResourceChangeDetail structures that describes the changes that CloudFormation will make to the resource.

resourceChange_scope :: Lens' ResourceChange (Maybe [ResourceAttribute])

For the Modify action, indicates which resource attribute is triggering this update, such as a change in the resource attribute's Metadata, Properties, or Tags.

resourceChange_moduleInfo :: Lens' ResourceChange (Maybe ModuleInfo)

Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

resourceChange_logicalResourceId :: Lens' ResourceChange (Maybe Text)

The resource's logical ID, which is defined in the stack's template.

resourceChange_changeSetId :: Lens' ResourceChange (Maybe Text)

The change set ID of the nested change set.

resourceChange_action :: Lens' ResourceChange (Maybe ChangeAction)

The action that CloudFormation takes on the resource, such as Add (adds a new resource), Modify (changes a resource), Remove (deletes a resource), Import (imports a resource), or Dynamic (exact action for the resource cannot be determined).

resourceChange_replacement :: Lens' ResourceChange (Maybe Replacement)

For the Modify action, indicates whether CloudFormation will replace the resource by creating a new one and deleting the old one. This value depends on the value of the RequiresRecreation property in the ResourceTargetDefinition structure. For example, if the RequiresRecreation field is Always and the Evaluation field is Static, Replacement is True. If the RequiresRecreation field is Always and the Evaluation field is Dynamic, Replacement is Conditionally.

If you have multiple changes with different RequiresRecreation values, the Replacement value depends on the change with the most impact. A RequiresRecreation value of Always has the most impact, followed by Conditionally, and then Never.