Network.AWS.CloudFormation.Types.ResourceTargetDefinition

Description

Documentation

data ResourceTargetDefinition

The field that CloudFormation will change, such as the name of a resource's property, and whether the resource will be recreated.

See: newResourceTargetDefinition smart constructor.

Constructors

ResourceTargetDefinition'

Fields

attribute :: Maybe ResourceAttribute

Indicates which resource attribute is triggering this update, such as a change in the resource attribute's Metadata, Properties, or Tags.

requiresRecreation :: Maybe RequiresRecreation

If the Attribute value is Properties, indicates whether a change to this property causes the resource to be recreated. The value can be Never, Always, or Conditionally. To determine the conditions for a Conditionally recreation, see the update behavior for that property in the CloudFormation User Guide.

name :: Maybe Text

If the Attribute value is Properties, the name of the property. For all other attributes, the value is null.

newResourceTargetDefinition :: ResourceTargetDefinition

Create a value of ResourceTargetDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attribute:ResourceTargetDefinition', resourceTargetDefinition_attribute - Indicates which resource attribute is triggering this update, such as a change in the resource attribute's Metadata, Properties, or Tags.

$sel:requiresRecreation:ResourceTargetDefinition', resourceTargetDefinition_requiresRecreation - If the Attribute value is Properties, indicates whether a change to this property causes the resource to be recreated. The value can be Never, Always, or Conditionally. To determine the conditions for a Conditionally recreation, see the update behavior for that property in the CloudFormation User Guide.

$sel:name:ResourceTargetDefinition', resourceTargetDefinition_name - If the Attribute value is Properties, the name of the property. For all other attributes, the value is null.

resourceTargetDefinition_attribute :: Lens' ResourceTargetDefinition (Maybe ResourceAttribute)

Indicates which resource attribute is triggering this update, such as a change in the resource attribute's Metadata, Properties, or Tags.

resourceTargetDefinition_requiresRecreation :: Lens' ResourceTargetDefinition (Maybe RequiresRecreation)

If the Attribute value is Properties, indicates whether a change to this property causes the resource to be recreated. The value can be Never, Always, or Conditionally. To determine the conditions for a Conditionally recreation, see the update behavior for that property in the CloudFormation User Guide.

resourceTargetDefinition_name :: Lens' ResourceTargetDefinition (Maybe Text)

If the Attribute value is Properties, the name of the property. For all other attributes, the value is null.