Network.AWS.CloudFormation.Types.ResourceToImport

Description

Documentation

data ResourceToImport

Describes the target resource of an import operation.

See: newResourceToImport smart constructor.

Constructors

ResourceToImport'

Fields

resourceType :: Text

The type of resource to import into your stack, such as AWS::S3::Bucket. For a list of supported resource types, see Resources that support import operations in the CloudFormation User Guide.

logicalResourceId :: Text

The logical ID of the target resource as specified in the template.

resourceIdentifier :: HashMap Text Text

A key-value pair that identifies the target resource. The key is an identifier property (for example, BucketName for AWS::S3::Bucket resources) and the value is the actual property value (for example, MyS3Bucket).

newResourceToImport

Arguments

:: Text
$sel:resourceType:ResourceToImport'

-> Text
$sel:logicalResourceId:ResourceToImport'

-> ResourceToImport

Create a value of ResourceToImport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ResourceToImport', resourceToImport_resourceType - The type of resource to import into your stack, such as AWS::S3::Bucket. For a list of supported resource types, see Resources that support import operations in the CloudFormation User Guide.

$sel:logicalResourceId:ResourceToImport', resourceToImport_logicalResourceId - The logical ID of the target resource as specified in the template.

$sel:resourceIdentifier:ResourceToImport', resourceToImport_resourceIdentifier - A key-value pair that identifies the target resource. The key is an identifier property (for example, BucketName for AWS::S3::Bucket resources) and the value is the actual property value (for example, MyS3Bucket).

resourceToImport_resourceType :: Lens' ResourceToImport Text

The type of resource to import into your stack, such as AWS::S3::Bucket. For a list of supported resource types, see Resources that support import operations in the CloudFormation User Guide.

resourceToImport_logicalResourceId :: Lens' ResourceToImport Text

The logical ID of the target resource as specified in the template.

resourceToImport_resourceIdentifier :: Lens' ResourceToImport (HashMap Text Text)

A key-value pair that identifies the target resource. The key is an identifier property (for example, BucketName for AWS::S3::Bucket resources) and the value is the actual property value (for example, MyS3Bucket).