Network.AWS.CloudFormation.Types.StackEvent

Description

Documentation

data StackEvent

The StackEvent data type.

See: newStackEvent smart constructor.

Constructors

StackEvent'

Fields

logicalResourceId :: Maybe Text

The logical name of the resource specified in the template.

physicalResourceId :: Maybe Text

The name or unique identifier associated with the physical instance of the resource.

resourceType :: Maybe Text

Type of resource. (For more information, go to Amazon Web Services Resource Types Reference in the CloudFormation User Guide.)

resourceStatusReason :: Maybe Text

Success/failure message associated with the resource.

resourceProperties :: Maybe Text

BLOB of the properties used to create the resource.

resourceStatus :: Maybe ResourceStatus

Current status of the resource.

clientRequestToken :: Maybe Text

The token passed to the operation that generated this event.

All events triggered by a given stack operation are assigned the same client request token, which you can use to track operations. For example, if you execute a CreateStack operation with the token token1, then all the StackEvents generated by that operation will have ClientRequestToken set as token1.

In the console, stack operations display the client request token on the Events tab. Stack operations that are initiated from the console use the token format Console-StackOperation-ID, which helps you easily identify the stack operation . For example, if you create a stack using the console, each stack event would be assigned the same token in the following format: Console-CreateStack-7f59c3cf-00d2-40c7-b2ff-e75db0987002.

stackId :: Text

The unique ID name of the instance of the stack.

eventId :: Text

The unique ID of this event.

stackName :: Text

The name associated with a stack.

timestamp :: ISO8601

Time the status was updated.

newStackEvent

Arguments

:: Text
$sel:stackId:StackEvent'

-> Text
$sel:eventId:StackEvent'

-> Text
$sel:stackName:StackEvent'

-> UTCTime
$sel:timestamp:StackEvent'

-> StackEvent

Create a value of StackEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logicalResourceId:StackEvent', stackEvent_logicalResourceId - The logical name of the resource specified in the template.

$sel:physicalResourceId:StackEvent', stackEvent_physicalResourceId - The name or unique identifier associated with the physical instance of the resource.

$sel:resourceType:StackEvent', stackEvent_resourceType - Type of resource. (For more information, go to Amazon Web Services Resource Types Reference in the CloudFormation User Guide.)

$sel:resourceStatusReason:StackEvent', stackEvent_resourceStatusReason - Success/failure message associated with the resource.

$sel:resourceProperties:StackEvent', stackEvent_resourceProperties - BLOB of the properties used to create the resource.

$sel:resourceStatus:StackEvent', stackEvent_resourceStatus - Current status of the resource.

$sel:clientRequestToken:StackEvent', stackEvent_clientRequestToken - The token passed to the operation that generated this event.

All events triggered by a given stack operation are assigned the same client request token, which you can use to track operations. For example, if you execute a CreateStack operation with the token token1, then all the StackEvents generated by that operation will have ClientRequestToken set as token1.

In the console, stack operations display the client request token on the Events tab. Stack operations that are initiated from the console use the token format Console-StackOperation-ID, which helps you easily identify the stack operation . For example, if you create a stack using the console, each stack event would be assigned the same token in the following format: Console-CreateStack-7f59c3cf-00d2-40c7-b2ff-e75db0987002.

$sel:stackId:StackEvent', stackEvent_stackId - The unique ID name of the instance of the stack.

$sel:eventId:StackEvent', stackEvent_eventId - The unique ID of this event.

$sel:stackName:StackEvent', stackEvent_stackName - The name associated with a stack.

$sel:timestamp:StackEvent', stackEvent_timestamp - Time the status was updated.

stackEvent_logicalResourceId :: Lens' StackEvent (Maybe Text)

The logical name of the resource specified in the template.

stackEvent_physicalResourceId :: Lens' StackEvent (Maybe Text)

The name or unique identifier associated with the physical instance of the resource.

stackEvent_resourceType :: Lens' StackEvent (Maybe Text)

Type of resource. (For more information, go to Amazon Web Services Resource Types Reference in the CloudFormation User Guide.)

stackEvent_resourceStatusReason :: Lens' StackEvent (Maybe Text)

Success/failure message associated with the resource.

stackEvent_resourceProperties :: Lens' StackEvent (Maybe Text)

BLOB of the properties used to create the resource.

stackEvent_resourceStatus :: Lens' StackEvent (Maybe ResourceStatus)

Current status of the resource.

stackEvent_clientRequestToken :: Lens' StackEvent (Maybe Text)

The token passed to the operation that generated this event.

All events triggered by a given stack operation are assigned the same client request token, which you can use to track operations. For example, if you execute a CreateStack operation with the token token1, then all the StackEvents generated by that operation will have ClientRequestToken set as token1.

In the console, stack operations display the client request token on the Events tab. Stack operations that are initiated from the console use the token format Console-StackOperation-ID, which helps you easily identify the stack operation . For example, if you create a stack using the console, each stack event would be assigned the same token in the following format: Console-CreateStack-7f59c3cf-00d2-40c7-b2ff-e75db0987002.

stackEvent_stackId :: Lens' StackEvent Text

The unique ID name of the instance of the stack.

stackEvent_eventId :: Lens' StackEvent Text

The unique ID of this event.

stackEvent_stackName :: Lens' StackEvent Text

The name associated with a stack.

stackEvent_timestamp :: Lens' StackEvent UTCTime

Time the status was updated.