Network.AWS.CloudFormation.Types.StackInstanceSummary

Description

Documentation

data StackInstanceSummary

The structure that contains summary information about a stack instance.

See: newStackInstanceSummary smart constructor.

Constructors

StackInstanceSummary'

Fields

status :: Maybe StackInstanceStatus

The status of the stack instance, in terms of its synchronization with its associated stack set.

INOPERABLE: A DeleteStackInstances operation has failed and left the stack in an unstable state. Stacks in this state are excluded from further UpdateStackSet operations. You might need to perform a DeleteStackInstances operation, with RetainStacks set to true, to delete the stack instance, and then delete the stack manually.

OUTDATED: The stack isn't currently up to date with the stack set because:

The associated stack failed during a CreateStackSet or UpdateStackSet operation.

The stack was part of a CreateStackSet or UpdateStackSet operation that failed or was stopped before the stack was created or updated.

CURRENT: The stack is currently up to date with the stack set.

stackId :: Maybe Text

The ID of the stack instance.

stackInstanceStatus :: Maybe StackInstanceComprehensiveStatus

The detailed status of the stack instance.

organizationalUnitId :: Maybe Text

Service-managed permissions

The organization root ID or organizational unit (OU) IDs that you specified for DeploymentTargets.

lastDriftCheckTimestamp :: Maybe ISO8601

Most recent time when CloudFormation performed a drift detection operation on the stack instance. This value will be NULL for any stack instance on which drift detection has not yet been performed.

stackSetId :: Maybe Text

The name or unique ID of the stack set that the stack instance is associated with.

driftStatus :: Maybe StackDriftStatus

Status of the stack instance's actual configuration compared to the expected template and parameter configuration of the stack set to which it belongs.

DRIFTED: The stack differs from the expected template and parameter configuration of the stack set to which it belongs. A stack instance is considered to have drifted if one or more of the resources in the associated stack have drifted.

NOT_CHECKED: CloudFormation has not checked if the stack instance differs from its expected stack set configuration.

IN_SYNC: The stack instance's actual configuration matches its expected stack set configuration.

UNKNOWN: This value is reserved for future use.

account :: Maybe Text

Self-managed permissions

The name of the Amazon Web Services account that the stack instance is associated with.

region :: Maybe Text

The name of the Amazon Web Services Region that the stack instance is associated with.

statusReason :: Maybe Text

The explanation for the specific status code assigned to this stack instance.

newStackInstanceSummary :: StackInstanceSummary

Create a value of StackInstanceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StackInstanceSummary', stackInstanceSummary_status - The status of the stack instance, in terms of its synchronization with its associated stack set.

INOPERABLE: A DeleteStackInstances operation has failed and left the stack in an unstable state. Stacks in this state are excluded from further UpdateStackSet operations. You might need to perform a DeleteStackInstances operation, with RetainStacks set to true, to delete the stack instance, and then delete the stack manually.

OUTDATED: The stack isn't currently up to date with the stack set because:

The associated stack failed during a CreateStackSet or UpdateStackSet operation.

The stack was part of a CreateStackSet or UpdateStackSet operation that failed or was stopped before the stack was created or updated.

CURRENT: The stack is currently up to date with the stack set.

$sel:stackId:StackInstanceSummary', stackInstanceSummary_stackId - The ID of the stack instance.

$sel:stackInstanceStatus:StackInstanceSummary', stackInstanceSummary_stackInstanceStatus - The detailed status of the stack instance.

$sel:organizationalUnitId:StackInstanceSummary', stackInstanceSummary_organizationalUnitId - [Service-managed permissions] The organization root ID or organizational unit (OU) IDs that you specified for DeploymentTargets.

$sel:lastDriftCheckTimestamp:StackInstanceSummary', stackInstanceSummary_lastDriftCheckTimestamp - Most recent time when CloudFormation performed a drift detection operation on the stack instance. This value will be NULL for any stack instance on which drift detection has not yet been performed.

$sel:stackSetId:StackInstanceSummary', stackInstanceSummary_stackSetId - The name or unique ID of the stack set that the stack instance is associated with.

$sel:driftStatus:StackInstanceSummary', stackInstanceSummary_driftStatus - Status of the stack instance's actual configuration compared to the expected template and parameter configuration of the stack set to which it belongs.

DRIFTED: The stack differs from the expected template and parameter configuration of the stack set to which it belongs. A stack instance is considered to have drifted if one or more of the resources in the associated stack have drifted.

NOT_CHECKED: CloudFormation has not checked if the stack instance differs from its expected stack set configuration.

IN_SYNC: The stack instance's actual configuration matches its expected stack set configuration.

UNKNOWN: This value is reserved for future use.

$sel:account:StackInstanceSummary', stackInstanceSummary_account - [Self-managed permissions] The name of the Amazon Web Services account that the stack instance is associated with.

$sel:region:StackInstanceSummary', stackInstanceSummary_region - The name of the Amazon Web Services Region that the stack instance is associated with.

$sel:statusReason:StackInstanceSummary', stackInstanceSummary_statusReason - The explanation for the specific status code assigned to this stack instance.

stackInstanceSummary_status :: Lens' StackInstanceSummary (Maybe StackInstanceStatus)

The status of the stack instance, in terms of its synchronization with its associated stack set.

INOPERABLE: A DeleteStackInstances operation has failed and left the stack in an unstable state. Stacks in this state are excluded from further UpdateStackSet operations. You might need to perform a DeleteStackInstances operation, with RetainStacks set to true, to delete the stack instance, and then delete the stack manually.

OUTDATED: The stack isn't currently up to date with the stack set because:

The associated stack failed during a CreateStackSet or UpdateStackSet operation.

The stack was part of a CreateStackSet or UpdateStackSet operation that failed or was stopped before the stack was created or updated.

CURRENT: The stack is currently up to date with the stack set.

stackInstanceSummary_stackId :: Lens' StackInstanceSummary (Maybe Text)

The ID of the stack instance.

stackInstanceSummary_stackInstanceStatus :: Lens' StackInstanceSummary (Maybe StackInstanceComprehensiveStatus)

The detailed status of the stack instance.

stackInstanceSummary_organizationalUnitId :: Lens' StackInstanceSummary (Maybe Text)

Service-managed permissions

The organization root ID or organizational unit (OU) IDs that you specified for DeploymentTargets.

stackInstanceSummary_lastDriftCheckTimestamp :: Lens' StackInstanceSummary (Maybe UTCTime)

Most recent time when CloudFormation performed a drift detection operation on the stack instance. This value will be NULL for any stack instance on which drift detection has not yet been performed.

stackInstanceSummary_stackSetId :: Lens' StackInstanceSummary (Maybe Text)

The name or unique ID of the stack set that the stack instance is associated with.

stackInstanceSummary_driftStatus :: Lens' StackInstanceSummary (Maybe StackDriftStatus)

Status of the stack instance's actual configuration compared to the expected template and parameter configuration of the stack set to which it belongs.

DRIFTED: The stack differs from the expected template and parameter configuration of the stack set to which it belongs. A stack instance is considered to have drifted if one or more of the resources in the associated stack have drifted.

NOT_CHECKED: CloudFormation has not checked if the stack instance differs from its expected stack set configuration.

IN_SYNC: The stack instance's actual configuration matches its expected stack set configuration.

UNKNOWN: This value is reserved for future use.

stackInstanceSummary_account :: Lens' StackInstanceSummary (Maybe Text)

Self-managed permissions

The name of the Amazon Web Services account that the stack instance is associated with.

stackInstanceSummary_region :: Lens' StackInstanceSummary (Maybe Text)

The name of the Amazon Web Services Region that the stack instance is associated with.

stackInstanceSummary_statusReason :: Lens' StackInstanceSummary (Maybe Text)

The explanation for the specific status code assigned to this stack instance.