Network.AWS.CloudFormation.Types.StackResource

Description

Documentation

data StackResource

The StackResource data type.

See: newStackResource smart constructor.

Constructors

StackResource'

Fields

physicalResourceId :: Maybe Text

The name or unique identifier that corresponds to a physical instance ID of a resource supported by CloudFormation.

resourceStatusReason :: Maybe Text

Success/failure message associated with the resource.

driftInformation :: Maybe StackResourceDriftInformation

Information about whether the resource's actual configuration differs, or has drifted, from its expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

moduleInfo :: Maybe ModuleInfo

Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

stackId :: Maybe Text

Unique identifier of the stack.

description :: Maybe Text

User defined description associated with the resource.

stackName :: Maybe Text

The name associated with the stack.

logicalResourceId :: Text

The logical name of the resource specified in the template.

resourceType :: Text

Type of resource. (For more information, go to Amazon Web Services Resource Types Reference in the CloudFormation User Guide.)

timestamp :: ISO8601

Time the status was updated.

resourceStatus :: ResourceStatus

Current status of the resource.

newStackResource

Arguments

:: Text
$sel:logicalResourceId:StackResource'

-> Text
$sel:resourceType:StackResource'

-> UTCTime
$sel:timestamp:StackResource'

-> ResourceStatus
$sel:resourceStatus:StackResource'

-> StackResource

Create a value of StackResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:physicalResourceId:StackResource', stackResource_physicalResourceId - The name or unique identifier that corresponds to a physical instance ID of a resource supported by CloudFormation.

$sel:resourceStatusReason:StackResource', stackResource_resourceStatusReason - Success/failure message associated with the resource.

$sel:driftInformation:StackResource', stackResource_driftInformation - Information about whether the resource's actual configuration differs, or has drifted, from its expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

$sel:moduleInfo:StackResource', stackResource_moduleInfo - Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

$sel:stackId:StackResource', stackResource_stackId - Unique identifier of the stack.

$sel:description:StackResource', stackResource_description - User defined description associated with the resource.

$sel:stackName:StackResource', stackResource_stackName - The name associated with the stack.

$sel:logicalResourceId:StackResource', stackResource_logicalResourceId - The logical name of the resource specified in the template.

$sel:resourceType:StackResource', stackResource_resourceType - Type of resource. (For more information, go to Amazon Web Services Resource Types Reference in the CloudFormation User Guide.)

$sel:timestamp:StackResource', stackResource_timestamp - Time the status was updated.

$sel:resourceStatus:StackResource', stackResource_resourceStatus - Current status of the resource.

stackResource_physicalResourceId :: Lens' StackResource (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of a resource supported by CloudFormation.

stackResource_resourceStatusReason :: Lens' StackResource (Maybe Text)

Success/failure message associated with the resource.

stackResource_driftInformation :: Lens' StackResource (Maybe StackResourceDriftInformation)

Information about whether the resource's actual configuration differs, or has drifted, from its expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

stackResource_moduleInfo :: Lens' StackResource (Maybe ModuleInfo)

Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

stackResource_stackId :: Lens' StackResource (Maybe Text)

Unique identifier of the stack.

stackResource_description :: Lens' StackResource (Maybe Text)

User defined description associated with the resource.

stackResource_stackName :: Lens' StackResource (Maybe Text)

The name associated with the stack.

stackResource_logicalResourceId :: Lens' StackResource Text

The logical name of the resource specified in the template.

stackResource_resourceType :: Lens' StackResource Text

Type of resource. (For more information, go to Amazon Web Services Resource Types Reference in the CloudFormation User Guide.)

stackResource_timestamp :: Lens' StackResource UTCTime

Time the status was updated.

stackResource_resourceStatus :: Lens' StackResource ResourceStatus

Current status of the resource.