Network.AWS.CloudFormation.Types.StackResourceDetail

Description

Documentation

data StackResourceDetail

Contains detailed information about the specified stack resource.

See: newStackResourceDetail smart constructor.

Constructors

StackResourceDetail'

Fields

physicalResourceId :: Maybe Text

The name or unique identifier that corresponds to a physical instance ID of a resource supported by CloudFormation.

resourceStatusReason :: Maybe Text

Success/failure message associated with the resource.

driftInformation :: Maybe StackResourceDriftInformation

Information about whether the resource's actual configuration differs, or has drifted, from its expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

moduleInfo :: Maybe ModuleInfo

Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

metadata :: Maybe Text

The content of the Metadata attribute declared for the resource. For more information, see Metadata Attribute in the CloudFormation User Guide.

stackId :: Maybe Text

Unique identifier of the stack.

description :: Maybe Text

User defined description associated with the resource.

stackName :: Maybe Text

The name associated with the stack.

logicalResourceId :: Text

The logical name of the resource specified in the template.

resourceType :: Text

Type of resource. ((For more information, go to Amazon Web Services Resource Types Reference in the CloudFormation User Guide.)

lastUpdatedTimestamp :: ISO8601

Time the status was updated.

resourceStatus :: ResourceStatus

Current status of the resource.

newStackResourceDetail

Arguments

:: Text
$sel:logicalResourceId:StackResourceDetail'

-> Text
$sel:resourceType:StackResourceDetail'

-> UTCTime
$sel:lastUpdatedTimestamp:StackResourceDetail'

-> ResourceStatus
$sel:resourceStatus:StackResourceDetail'

-> StackResourceDetail

Create a value of StackResourceDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:physicalResourceId:StackResourceDetail', stackResourceDetail_physicalResourceId - The name or unique identifier that corresponds to a physical instance ID of a resource supported by CloudFormation.

$sel:resourceStatusReason:StackResourceDetail', stackResourceDetail_resourceStatusReason - Success/failure message associated with the resource.

$sel:driftInformation:StackResourceDetail', stackResourceDetail_driftInformation - Information about whether the resource's actual configuration differs, or has drifted, from its expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

$sel:moduleInfo:StackResourceDetail', stackResourceDetail_moduleInfo - Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

$sel:metadata:StackResourceDetail', stackResourceDetail_metadata - The content of the Metadata attribute declared for the resource. For more information, see Metadata Attribute in the CloudFormation User Guide.

$sel:stackId:StackResourceDetail', stackResourceDetail_stackId - Unique identifier of the stack.

$sel:description:StackResourceDetail', stackResourceDetail_description - User defined description associated with the resource.

$sel:stackName:StackResourceDetail', stackResourceDetail_stackName - The name associated with the stack.

$sel:logicalResourceId:StackResourceDetail', stackResourceDetail_logicalResourceId - The logical name of the resource specified in the template.

$sel:resourceType:StackResourceDetail', stackResourceDetail_resourceType - Type of resource. ((For more information, go to Amazon Web Services Resource Types Reference in the CloudFormation User Guide.)

$sel:lastUpdatedTimestamp:StackResourceDetail', stackResourceDetail_lastUpdatedTimestamp - Time the status was updated.

$sel:resourceStatus:StackResourceDetail', stackResourceDetail_resourceStatus - Current status of the resource.

stackResourceDetail_physicalResourceId :: Lens' StackResourceDetail (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of a resource supported by CloudFormation.

stackResourceDetail_resourceStatusReason :: Lens' StackResourceDetail (Maybe Text)

Success/failure message associated with the resource.

stackResourceDetail_driftInformation :: Lens' StackResourceDetail (Maybe StackResourceDriftInformation)

Information about whether the resource's actual configuration differs, or has drifted, from its expected configuration, as defined in the stack template and any values specified as template parameters. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

stackResourceDetail_moduleInfo :: Lens' StackResourceDetail (Maybe ModuleInfo)

Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

stackResourceDetail_metadata :: Lens' StackResourceDetail (Maybe Text)

The content of the Metadata attribute declared for the resource. For more information, see Metadata Attribute in the CloudFormation User Guide.

stackResourceDetail_stackId :: Lens' StackResourceDetail (Maybe Text)

Unique identifier of the stack.

stackResourceDetail_description :: Lens' StackResourceDetail (Maybe Text)

User defined description associated with the resource.

stackResourceDetail_stackName :: Lens' StackResourceDetail (Maybe Text)

The name associated with the stack.

stackResourceDetail_logicalResourceId :: Lens' StackResourceDetail Text

The logical name of the resource specified in the template.

stackResourceDetail_resourceType :: Lens' StackResourceDetail Text

Type of resource. ((For more information, go to Amazon Web Services Resource Types Reference in the CloudFormation User Guide.)

stackResourceDetail_lastUpdatedTimestamp :: Lens' StackResourceDetail UTCTime

Time the status was updated.

stackResourceDetail_resourceStatus :: Lens' StackResourceDetail ResourceStatus

Current status of the resource.