Network.AWS.CloudFormation.Types.StackResourceDrift

Description

Documentation

data StackResourceDrift

Contains the drift information for a resource that has been checked for drift. This includes actual and expected property values for resources in which CloudFormation has detected drift. Only resource properties explicitly defined in the stack template are checked for drift. For more information, see Detecting Unregulated Configuration Changes to Stacks and Resources.

Resources that do not currently support drift detection cannot be checked. For a list of resources that support drift detection, see Resources that Support Drift Detection.

Use DetectStackResourceDrift to detect drift on individual resources, or DetectStackDrift to detect drift on all resources in a given stack that support drift detection.

See: newStackResourceDrift smart constructor.

Constructors

StackResourceDrift'

Fields

actualProperties :: Maybe Text

A JSON structure containing the actual property values of the stack resource.

For resources whose StackResourceDriftStatus is DELETED, this structure will not be present.

physicalResourceIdContext :: Maybe [PhysicalResourceIdContextKeyValuePair]

Context information that enables CloudFormation to uniquely identify a resource. CloudFormation uses context key-value pairs in cases where a resource's logical and physical IDs are not enough to uniquely identify that resource. Each context key-value pair specifies a unique resource that contains the targeted resource.

physicalResourceId :: Maybe Text

The name or unique identifier that corresponds to a physical instance ID of a resource supported by CloudFormation.

expectedProperties :: Maybe Text

A JSON structure containing the expected property values of the stack resource, as defined in the stack template and any values specified as template parameters.

For resources whose StackResourceDriftStatus is DELETED, this structure will not be present.

moduleInfo :: Maybe ModuleInfo

Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

propertyDifferences :: Maybe [PropertyDifference]

A collection of the resource properties whose actual values differ from their expected values. These will be present only for resources whose StackResourceDriftStatus is MODIFIED.

stackId :: Text

The ID of the stack.

logicalResourceId :: Text

The logical name of the resource specified in the template.

resourceType :: Text

The type of the resource.

stackResourceDriftStatus :: StackResourceDriftStatus

Status of the resource's actual configuration compared to its expected configuration

DELETED: The resource differs from its expected template configuration because the resource has been deleted.

MODIFIED: One or more resource properties differ from their expected values (as defined in the stack template and any values specified as template parameters).

IN_SYNC: The resources's actual configuration matches its expected template configuration.

NOT_CHECKED: CloudFormation does not currently return this value.

timestamp :: ISO8601

Time at which CloudFormation performed drift detection on the stack resource.

newStackResourceDrift

Arguments

:: Text
$sel:stackId:StackResourceDrift'

-> Text
$sel:logicalResourceId:StackResourceDrift'

-> Text
$sel:resourceType:StackResourceDrift'

-> StackResourceDriftStatus
$sel:stackResourceDriftStatus:StackResourceDrift'

-> UTCTime
$sel:timestamp:StackResourceDrift'

-> StackResourceDrift

Create a value of StackResourceDrift with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actualProperties:StackResourceDrift', stackResourceDrift_actualProperties - A JSON structure containing the actual property values of the stack resource.

For resources whose StackResourceDriftStatus is DELETED, this structure will not be present.

$sel:physicalResourceIdContext:StackResourceDrift', stackResourceDrift_physicalResourceIdContext - Context information that enables CloudFormation to uniquely identify a resource. CloudFormation uses context key-value pairs in cases where a resource's logical and physical IDs are not enough to uniquely identify that resource. Each context key-value pair specifies a unique resource that contains the targeted resource.

$sel:physicalResourceId:StackResourceDrift', stackResourceDrift_physicalResourceId - The name or unique identifier that corresponds to a physical instance ID of a resource supported by CloudFormation.

$sel:expectedProperties:StackResourceDrift', stackResourceDrift_expectedProperties - A JSON structure containing the expected property values of the stack resource, as defined in the stack template and any values specified as template parameters.

For resources whose StackResourceDriftStatus is DELETED, this structure will not be present.

$sel:moduleInfo:StackResourceDrift', stackResourceDrift_moduleInfo - Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

$sel:propertyDifferences:StackResourceDrift', stackResourceDrift_propertyDifferences - A collection of the resource properties whose actual values differ from their expected values. These will be present only for resources whose StackResourceDriftStatus is MODIFIED.

$sel:stackId:StackResourceDrift', stackResourceDrift_stackId - The ID of the stack.

$sel:logicalResourceId:StackResourceDrift', stackResourceDrift_logicalResourceId - The logical name of the resource specified in the template.

$sel:resourceType:StackResourceDrift', stackResourceDrift_resourceType - The type of the resource.

$sel:stackResourceDriftStatus:StackResourceDrift', stackResourceDrift_stackResourceDriftStatus - Status of the resource's actual configuration compared to its expected configuration

DELETED: The resource differs from its expected template configuration because the resource has been deleted.

MODIFIED: One or more resource properties differ from their expected values (as defined in the stack template and any values specified as template parameters).

IN_SYNC: The resources's actual configuration matches its expected template configuration.

NOT_CHECKED: CloudFormation does not currently return this value.

$sel:timestamp:StackResourceDrift', stackResourceDrift_timestamp - Time at which CloudFormation performed drift detection on the stack resource.

stackResourceDrift_actualProperties :: Lens' StackResourceDrift (Maybe Text)

A JSON structure containing the actual property values of the stack resource.

For resources whose StackResourceDriftStatus is DELETED, this structure will not be present.

stackResourceDrift_physicalResourceIdContext :: Lens' StackResourceDrift (Maybe [PhysicalResourceIdContextKeyValuePair])

Context information that enables CloudFormation to uniquely identify a resource. CloudFormation uses context key-value pairs in cases where a resource's logical and physical IDs are not enough to uniquely identify that resource. Each context key-value pair specifies a unique resource that contains the targeted resource.

stackResourceDrift_physicalResourceId :: Lens' StackResourceDrift (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of a resource supported by CloudFormation.

stackResourceDrift_expectedProperties :: Lens' StackResourceDrift (Maybe Text)

A JSON structure containing the expected property values of the stack resource, as defined in the stack template and any values specified as template parameters.

For resources whose StackResourceDriftStatus is DELETED, this structure will not be present.

stackResourceDrift_moduleInfo :: Lens' StackResourceDrift (Maybe ModuleInfo)

Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

stackResourceDrift_propertyDifferences :: Lens' StackResourceDrift (Maybe [PropertyDifference])

A collection of the resource properties whose actual values differ from their expected values. These will be present only for resources whose StackResourceDriftStatus is MODIFIED.

stackResourceDrift_stackId :: Lens' StackResourceDrift Text

The ID of the stack.

stackResourceDrift_logicalResourceId :: Lens' StackResourceDrift Text

The logical name of the resource specified in the template.

stackResourceDrift_resourceType :: Lens' StackResourceDrift Text

The type of the resource.

stackResourceDrift_stackResourceDriftStatus :: Lens' StackResourceDrift StackResourceDriftStatus

Status of the resource's actual configuration compared to its expected configuration

DELETED: The resource differs from its expected template configuration because the resource has been deleted.

MODIFIED: One or more resource properties differ from their expected values (as defined in the stack template and any values specified as template parameters).

IN_SYNC: The resources's actual configuration matches its expected template configuration.

NOT_CHECKED: CloudFormation does not currently return this value.

stackResourceDrift_timestamp :: Lens' StackResourceDrift UTCTime

Time at which CloudFormation performed drift detection on the stack resource.