Network.AWS.CloudFormation.Types.StackResourceDriftInformation

Description

Documentation

data StackResourceDriftInformation

Contains information about whether the resource's actual configuration differs, or has drifted, from its expected configuration.

See: newStackResourceDriftInformation smart constructor.

Constructors

StackResourceDriftInformation'

Fields

lastCheckTimestamp :: Maybe ISO8601

When CloudFormation last checked if the resource had drifted from its expected configuration.

stackResourceDriftStatus :: StackResourceDriftStatus

Status of the resource's actual configuration compared to its expected configuration

DELETED: The resource differs from its expected configuration in that it has been deleted.

MODIFIED: The resource differs from its expected configuration.

NOT_CHECKED: CloudFormation has not checked if the resource differs from its expected configuration.

Any resources that do not currently support drift detection have a status of NOT_CHECKED. For more information, see Resources that Support Drift Detection.

IN_SYNC: The resources's actual configuration matches its expected configuration.

newStackResourceDriftInformation

Arguments

:: StackResourceDriftStatus
$sel:stackResourceDriftStatus:StackResourceDriftInformation'

-> StackResourceDriftInformation

Create a value of StackResourceDriftInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastCheckTimestamp:StackResourceDriftInformation', stackResourceDriftInformation_lastCheckTimestamp - When CloudFormation last checked if the resource had drifted from its expected configuration.

$sel:stackResourceDriftStatus:StackResourceDriftInformation', stackResourceDriftInformation_stackResourceDriftStatus - Status of the resource's actual configuration compared to its expected configuration

DELETED: The resource differs from its expected configuration in that it has been deleted.

MODIFIED: The resource differs from its expected configuration.

NOT_CHECKED: CloudFormation has not checked if the resource differs from its expected configuration.

Any resources that do not currently support drift detection have a status of NOT_CHECKED. For more information, see Resources that Support Drift Detection.

IN_SYNC: The resources's actual configuration matches its expected configuration.

stackResourceDriftInformation_lastCheckTimestamp :: Lens' StackResourceDriftInformation (Maybe UTCTime)

When CloudFormation last checked if the resource had drifted from its expected configuration.

stackResourceDriftInformation_stackResourceDriftStatus :: Lens' StackResourceDriftInformation StackResourceDriftStatus

Status of the resource's actual configuration compared to its expected configuration

DELETED: The resource differs from its expected configuration in that it has been deleted.

MODIFIED: The resource differs from its expected configuration.

NOT_CHECKED: CloudFormation has not checked if the resource differs from its expected configuration.

Any resources that do not currently support drift detection have a status of NOT_CHECKED. For more information, see Resources that Support Drift Detection.

IN_SYNC: The resources's actual configuration matches its expected configuration.