Network.AWS.CloudFormation.Types.StackResourceDriftInformationSummary

Description

Documentation

data StackResourceDriftInformationSummary

Summarizes information about whether the resource's actual configuration differs, or has drifted, from its expected configuration.

See: newStackResourceDriftInformationSummary smart constructor.

Constructors

StackResourceDriftInformationSummary'

Fields

lastCheckTimestamp :: Maybe ISO8601

When CloudFormation last checked if the resource had drifted from its expected configuration.

stackResourceDriftStatus :: StackResourceDriftStatus

Status of the resource's actual configuration compared to its expected configuration

DELETED: The resource differs from its expected configuration in that it has been deleted.

MODIFIED: The resource differs from its expected configuration.

NOT_CHECKED: CloudFormation has not checked if the resource differs from its expected configuration.

Any resources that do not currently support drift detection have a status of NOT_CHECKED. For more information, see Resources that Support Drift Detection. If you performed an ContinueUpdateRollback operation on a stack, any resources included in ResourcesToSkip will also have a status of NOT_CHECKED. For more information on skipping resources during rollback operations, see Continue Rolling Back an Update in the CloudFormation User Guide.

IN_SYNC: The resources's actual configuration matches its expected configuration.

newStackResourceDriftInformationSummary

Arguments

:: StackResourceDriftStatus
$sel:stackResourceDriftStatus:StackResourceDriftInformationSummary'

-> StackResourceDriftInformationSummary

Create a value of StackResourceDriftInformationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastCheckTimestamp:StackResourceDriftInformationSummary', stackResourceDriftInformationSummary_lastCheckTimestamp - When CloudFormation last checked if the resource had drifted from its expected configuration.

$sel:stackResourceDriftStatus:StackResourceDriftInformationSummary', stackResourceDriftInformationSummary_stackResourceDriftStatus - Status of the resource's actual configuration compared to its expected configuration

DELETED: The resource differs from its expected configuration in that it has been deleted.

MODIFIED: The resource differs from its expected configuration.

NOT_CHECKED: CloudFormation has not checked if the resource differs from its expected configuration.

Any resources that do not currently support drift detection have a status of NOT_CHECKED. For more information, see Resources that Support Drift Detection. If you performed an ContinueUpdateRollback operation on a stack, any resources included in ResourcesToSkip will also have a status of NOT_CHECKED. For more information on skipping resources during rollback operations, see Continue Rolling Back an Update in the CloudFormation User Guide.

IN_SYNC: The resources's actual configuration matches its expected configuration.

stackResourceDriftInformationSummary_lastCheckTimestamp :: Lens' StackResourceDriftInformationSummary (Maybe UTCTime)

When CloudFormation last checked if the resource had drifted from its expected configuration.

stackResourceDriftInformationSummary_stackResourceDriftStatus :: Lens' StackResourceDriftInformationSummary StackResourceDriftStatus

Status of the resource's actual configuration compared to its expected configuration

DELETED: The resource differs from its expected configuration in that it has been deleted.

MODIFIED: The resource differs from its expected configuration.

NOT_CHECKED: CloudFormation has not checked if the resource differs from its expected configuration.

Any resources that do not currently support drift detection have a status of NOT_CHECKED. For more information, see Resources that Support Drift Detection. If you performed an ContinueUpdateRollback operation on a stack, any resources included in ResourcesToSkip will also have a status of NOT_CHECKED. For more information on skipping resources during rollback operations, see Continue Rolling Back an Update in the CloudFormation User Guide.

IN_SYNC: The resources's actual configuration matches its expected configuration.