Network.AWS.CloudFormation.Types.StackSetDriftDetectionDetails

Description

Documentation

data StackSetDriftDetectionDetails

Detailed information about the drift status of the stack set.

For stack sets, contains information about the last completed drift operation performed on the stack set. Information about drift operations in-progress is not included.

For stack set operations, includes information about drift operations currently being performed on the stack set.

For more information, see Detecting Unmanaged Changes in Stack Sets in the CloudFormation User Guide.

See: newStackSetDriftDetectionDetails smart constructor.

Constructors

StackSetDriftDetectionDetails'

Fields

lastDriftCheckTimestamp :: Maybe ISO8601

Most recent time when CloudFormation performed a drift detection operation on the stack set. This value will be NULL for any stack set on which drift detection has not yet been performed.

totalStackInstancesCount :: Maybe Natural

The total number of stack instances belonging to this stack set.

The total number of stack instances is equal to the total of:

Stack instances that match the stack set configuration.

Stack instances that have drifted from the stack set configuration.

Stack instances where the drift detection operation has failed.

Stack instances currently being checked for drift.

inProgressStackInstancesCount :: Maybe Natural

The number of stack instances that are currently being checked for drift.

driftedStackInstancesCount :: Maybe Natural

The number of stack instances that have drifted from the expected template and parameter configuration of the stack set. A stack instance is considered to have drifted if one or more of the resources in the associated stack do not match their expected configuration.

driftDetectionStatus :: Maybe StackSetDriftDetectionStatus

The status of the stack set drift detection operation.

COMPLETED: The drift detection operation completed without failing on any stack instances.

FAILED: The drift detection operation exceeded the specified failure tolerance.

PARTIAL_SUCCESS: The drift detection operation completed without exceeding the failure tolerance for the operation.

IN_PROGRESS: The drift detection operation is currently being performed.

STOPPED: The user has cancelled the drift detection operation.

driftStatus :: Maybe StackSetDriftStatus

Status of the stack set's actual configuration compared to its expected template and parameter configuration. A stack set is considered to have drifted if one or more of its stack instances have drifted from their expected template and parameter configuration.

DRIFTED: One or more of the stack instances belonging to the stack set stack differs from the expected template and parameter configuration. A stack instance is considered to have drifted if one or more of the resources in the associated stack have drifted.

NOT_CHECKED: CloudFormation has not checked the stack set for drift.

IN_SYNC: All of the stack instances belonging to the stack set stack match from the expected template and parameter configuration.

failedStackInstancesCount :: Maybe Natural

The number of stack instances for which the drift detection operation failed.

inSyncStackInstancesCount :: Maybe Natural

The number of stack instances which match the expected template and parameter configuration of the stack set.

newStackSetDriftDetectionDetails :: StackSetDriftDetectionDetails

Create a value of StackSetDriftDetectionDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastDriftCheckTimestamp:StackSetDriftDetectionDetails', stackSetDriftDetectionDetails_lastDriftCheckTimestamp - Most recent time when CloudFormation performed a drift detection operation on the stack set. This value will be NULL for any stack set on which drift detection has not yet been performed.

$sel:totalStackInstancesCount:StackSetDriftDetectionDetails', stackSetDriftDetectionDetails_totalStackInstancesCount - The total number of stack instances belonging to this stack set.

The total number of stack instances is equal to the total of:

Stack instances that match the stack set configuration.

Stack instances that have drifted from the stack set configuration.

Stack instances where the drift detection operation has failed.

Stack instances currently being checked for drift.

$sel:inProgressStackInstancesCount:StackSetDriftDetectionDetails', stackSetDriftDetectionDetails_inProgressStackInstancesCount - The number of stack instances that are currently being checked for drift.

$sel:driftedStackInstancesCount:StackSetDriftDetectionDetails', stackSetDriftDetectionDetails_driftedStackInstancesCount - The number of stack instances that have drifted from the expected template and parameter configuration of the stack set. A stack instance is considered to have drifted if one or more of the resources in the associated stack do not match their expected configuration.

$sel:driftDetectionStatus:StackSetDriftDetectionDetails', stackSetDriftDetectionDetails_driftDetectionStatus - The status of the stack set drift detection operation.

COMPLETED: The drift detection operation completed without failing on any stack instances.

FAILED: The drift detection operation exceeded the specified failure tolerance.

PARTIAL_SUCCESS: The drift detection operation completed without exceeding the failure tolerance for the operation.

IN_PROGRESS: The drift detection operation is currently being performed.

STOPPED: The user has cancelled the drift detection operation.

$sel:driftStatus:StackSetDriftDetectionDetails', stackSetDriftDetectionDetails_driftStatus - Status of the stack set's actual configuration compared to its expected template and parameter configuration. A stack set is considered to have drifted if one or more of its stack instances have drifted from their expected template and parameter configuration.

DRIFTED: One or more of the stack instances belonging to the stack set stack differs from the expected template and parameter configuration. A stack instance is considered to have drifted if one or more of the resources in the associated stack have drifted.

NOT_CHECKED: CloudFormation has not checked the stack set for drift.

IN_SYNC: All of the stack instances belonging to the stack set stack match from the expected template and parameter configuration.

$sel:failedStackInstancesCount:StackSetDriftDetectionDetails', stackSetDriftDetectionDetails_failedStackInstancesCount - The number of stack instances for which the drift detection operation failed.

$sel:inSyncStackInstancesCount:StackSetDriftDetectionDetails', stackSetDriftDetectionDetails_inSyncStackInstancesCount - The number of stack instances which match the expected template and parameter configuration of the stack set.

stackSetDriftDetectionDetails_lastDriftCheckTimestamp :: Lens' StackSetDriftDetectionDetails (Maybe UTCTime)

Most recent time when CloudFormation performed a drift detection operation on the stack set. This value will be NULL for any stack set on which drift detection has not yet been performed.

stackSetDriftDetectionDetails_totalStackInstancesCount :: Lens' StackSetDriftDetectionDetails (Maybe Natural)

The total number of stack instances belonging to this stack set.

The total number of stack instances is equal to the total of:

Stack instances that match the stack set configuration.

Stack instances that have drifted from the stack set configuration.

Stack instances where the drift detection operation has failed.

Stack instances currently being checked for drift.

stackSetDriftDetectionDetails_inProgressStackInstancesCount :: Lens' StackSetDriftDetectionDetails (Maybe Natural)

The number of stack instances that are currently being checked for drift.

stackSetDriftDetectionDetails_driftedStackInstancesCount :: Lens' StackSetDriftDetectionDetails (Maybe Natural)

The number of stack instances that have drifted from the expected template and parameter configuration of the stack set. A stack instance is considered to have drifted if one or more of the resources in the associated stack do not match their expected configuration.

stackSetDriftDetectionDetails_driftDetectionStatus :: Lens' StackSetDriftDetectionDetails (Maybe StackSetDriftDetectionStatus)

The status of the stack set drift detection operation.

COMPLETED: The drift detection operation completed without failing on any stack instances.

FAILED: The drift detection operation exceeded the specified failure tolerance.

PARTIAL_SUCCESS: The drift detection operation completed without exceeding the failure tolerance for the operation.

IN_PROGRESS: The drift detection operation is currently being performed.

STOPPED: The user has cancelled the drift detection operation.

stackSetDriftDetectionDetails_driftStatus :: Lens' StackSetDriftDetectionDetails (Maybe StackSetDriftStatus)

Status of the stack set's actual configuration compared to its expected template and parameter configuration. A stack set is considered to have drifted if one or more of its stack instances have drifted from their expected template and parameter configuration.

DRIFTED: One or more of the stack instances belonging to the stack set stack differs from the expected template and parameter configuration. A stack instance is considered to have drifted if one or more of the resources in the associated stack have drifted.

NOT_CHECKED: CloudFormation has not checked the stack set for drift.

IN_SYNC: All of the stack instances belonging to the stack set stack match from the expected template and parameter configuration.

stackSetDriftDetectionDetails_failedStackInstancesCount :: Lens' StackSetDriftDetectionDetails (Maybe Natural)

The number of stack instances for which the drift detection operation failed.

stackSetDriftDetectionDetails_inSyncStackInstancesCount :: Lens' StackSetDriftDetectionDetails (Maybe Natural)

The number of stack instances which match the expected template and parameter configuration of the stack set.