Network.AWS.CloudFormation.Types.StackSetOperationResultSummary

Description

Documentation

data StackSetOperationResultSummary

The structure that contains information about a specified operation's results for a given account in a given Region.

See: newStackSetOperationResultSummary smart constructor.

Constructors

StackSetOperationResultSummary'

Fields

status :: Maybe StackSetOperationResultStatus

The result status of the stack set operation for the given account in the given Region.

CANCELLED: The operation in the specified account and Region has been cancelled. This is either because a user has stopped the stack set operation, or because the failure tolerance of the stack set operation has been exceeded.

FAILED: The operation in the specified account and Region failed.

If the stack set operation fails in enough accounts within a Region, the failure tolerance for the stack set operation as a whole might be exceeded.

RUNNING: The operation in the specified account and Region is currently in progress.

PENDING: The operation in the specified account and Region has yet to start.

SUCCEEDED: The operation in the specified account and Region completed successfully.

account :: Maybe Text

Self-managed permissions

The name of the Amazon Web Services account for this operation result.

accountGateResult :: Maybe AccountGateResult

The results of the account gate function CloudFormation invokes, if present, before proceeding with stack set operations in an account

organizationalUnitId :: Maybe Text

Service-managed permissions

The organization root ID or organizational unit (OU) IDs that you specified for DeploymentTargets.

region :: Maybe Text

The name of the Amazon Web Services Region for this operation result.

statusReason :: Maybe Text

The reason for the assigned result status.

newStackSetOperationResultSummary :: StackSetOperationResultSummary

Create a value of StackSetOperationResultSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StackSetOperationResultSummary', stackSetOperationResultSummary_status - The result status of the stack set operation for the given account in the given Region.

CANCELLED: The operation in the specified account and Region has been cancelled. This is either because a user has stopped the stack set operation, or because the failure tolerance of the stack set operation has been exceeded.

FAILED: The operation in the specified account and Region failed.

If the stack set operation fails in enough accounts within a Region, the failure tolerance for the stack set operation as a whole might be exceeded.

RUNNING: The operation in the specified account and Region is currently in progress.

PENDING: The operation in the specified account and Region has yet to start.

SUCCEEDED: The operation in the specified account and Region completed successfully.

$sel:account:StackSetOperationResultSummary', stackSetOperationResultSummary_account - [Self-managed permissions] The name of the Amazon Web Services account for this operation result.

$sel:accountGateResult:StackSetOperationResultSummary', stackSetOperationResultSummary_accountGateResult - The results of the account gate function CloudFormation invokes, if present, before proceeding with stack set operations in an account

$sel:organizationalUnitId:StackSetOperationResultSummary', stackSetOperationResultSummary_organizationalUnitId - [Service-managed permissions] The organization root ID or organizational unit (OU) IDs that you specified for DeploymentTargets.

$sel:region:StackSetOperationResultSummary', stackSetOperationResultSummary_region - The name of the Amazon Web Services Region for this operation result.

$sel:statusReason:StackSetOperationResultSummary', stackSetOperationResultSummary_statusReason - The reason for the assigned result status.

stackSetOperationResultSummary_status :: Lens' StackSetOperationResultSummary (Maybe StackSetOperationResultStatus)

The result status of the stack set operation for the given account in the given Region.

CANCELLED: The operation in the specified account and Region has been cancelled. This is either because a user has stopped the stack set operation, or because the failure tolerance of the stack set operation has been exceeded.

FAILED: The operation in the specified account and Region failed.

If the stack set operation fails in enough accounts within a Region, the failure tolerance for the stack set operation as a whole might be exceeded.

RUNNING: The operation in the specified account and Region is currently in progress.

PENDING: The operation in the specified account and Region has yet to start.

SUCCEEDED: The operation in the specified account and Region completed successfully.

stackSetOperationResultSummary_account :: Lens' StackSetOperationResultSummary (Maybe Text)

Self-managed permissions

The name of the Amazon Web Services account for this operation result.

stackSetOperationResultSummary_accountGateResult :: Lens' StackSetOperationResultSummary (Maybe AccountGateResult)

The results of the account gate function CloudFormation invokes, if present, before proceeding with stack set operations in an account

stackSetOperationResultSummary_organizationalUnitId :: Lens' StackSetOperationResultSummary (Maybe Text)

Service-managed permissions

The organization root ID or organizational unit (OU) IDs that you specified for DeploymentTargets.

stackSetOperationResultSummary_region :: Lens' StackSetOperationResultSummary (Maybe Text)

The name of the Amazon Web Services Region for this operation result.

stackSetOperationResultSummary_statusReason :: Lens' StackSetOperationResultSummary (Maybe Text)

The reason for the assigned result status.