Network.AWS.CloudFormation.Types.StackSetOperationSummary

Description

Documentation

data StackSetOperationSummary

The structures that contain summary information about the specified operation.

See: newStackSetOperationSummary smart constructor.

Constructors

StackSetOperationSummary'

Fields

creationTimestamp :: Maybe ISO8601

The time at which the operation was initiated. Note that the creation times for the stack set operation might differ from the creation time of the individual stacks themselves. This is because CloudFormation needs to perform preparatory work for the operation, such as dispatching the work to the requested Regions, before actually creating the first stacks.

status :: Maybe StackSetOperationStatus

The overall status of the operation.

FAILED: The operation exceeded the specified failure tolerance. The failure tolerance value that you've set for an operation is applied for each Region during stack create and update operations. If the number of failed stacks within a Region exceeds the failure tolerance, the status of the operation in the Region is set to FAILED. This in turn sets the status of the operation as a whole to FAILED, and CloudFormation cancels the operation in any remaining Regions.

QUEUED: [Service-managed permissions] For automatic deployments that require a sequence of operations, the operation is queued to be performed. For more information, see the stack set operation status codes in the CloudFormation User Guide.

RUNNING: The operation is currently being performed.

STOPPED: The user has cancelled the operation.

STOPPING: The operation is in the process of stopping, at user request.

SUCCEEDED: The operation completed creating or updating all the specified stacks without exceeding the failure tolerance for the operation.

endTimestamp :: Maybe ISO8601

The time at which the stack set operation ended, across all accounts and Regions specified. Note that this doesn't necessarily mean that the stack set operation was successful, or even attempted, in each account or Region.

operationId :: Maybe Text

The unique ID of the stack set operation.

action :: Maybe StackSetOperationAction

The type of operation: CREATE, UPDATE, or DELETE. Create and delete operations affect only the specified stack instances that are associated with the specified stack set. Update operations affect both the stack set itself as well as all associated stack set instances.

newStackSetOperationSummary :: StackSetOperationSummary

Create a value of StackSetOperationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:StackSetOperationSummary', stackSetOperationSummary_creationTimestamp - The time at which the operation was initiated. Note that the creation times for the stack set operation might differ from the creation time of the individual stacks themselves. This is because CloudFormation needs to perform preparatory work for the operation, such as dispatching the work to the requested Regions, before actually creating the first stacks.

$sel:status:StackSetOperationSummary', stackSetOperationSummary_status - The overall status of the operation.

FAILED: The operation exceeded the specified failure tolerance. The failure tolerance value that you've set for an operation is applied for each Region during stack create and update operations. If the number of failed stacks within a Region exceeds the failure tolerance, the status of the operation in the Region is set to FAILED. This in turn sets the status of the operation as a whole to FAILED, and CloudFormation cancels the operation in any remaining Regions.

QUEUED: [Service-managed permissions] For automatic deployments that require a sequence of operations, the operation is queued to be performed. For more information, see the stack set operation status codes in the CloudFormation User Guide.

RUNNING: The operation is currently being performed.

STOPPED: The user has cancelled the operation.

STOPPING: The operation is in the process of stopping, at user request.

SUCCEEDED: The operation completed creating or updating all the specified stacks without exceeding the failure tolerance for the operation.

$sel:endTimestamp:StackSetOperationSummary', stackSetOperationSummary_endTimestamp - The time at which the stack set operation ended, across all accounts and Regions specified. Note that this doesn't necessarily mean that the stack set operation was successful, or even attempted, in each account or Region.

$sel:operationId:StackSetOperationSummary', stackSetOperationSummary_operationId - The unique ID of the stack set operation.

$sel:action:StackSetOperationSummary', stackSetOperationSummary_action - The type of operation: CREATE, UPDATE, or DELETE. Create and delete operations affect only the specified stack instances that are associated with the specified stack set. Update operations affect both the stack set itself as well as all associated stack set instances.

stackSetOperationSummary_creationTimestamp :: Lens' StackSetOperationSummary (Maybe UTCTime)

The time at which the operation was initiated. Note that the creation times for the stack set operation might differ from the creation time of the individual stacks themselves. This is because CloudFormation needs to perform preparatory work for the operation, such as dispatching the work to the requested Regions, before actually creating the first stacks.

stackSetOperationSummary_status :: Lens' StackSetOperationSummary (Maybe StackSetOperationStatus)

The overall status of the operation.

FAILED: The operation exceeded the specified failure tolerance. The failure tolerance value that you've set for an operation is applied for each Region during stack create and update operations. If the number of failed stacks within a Region exceeds the failure tolerance, the status of the operation in the Region is set to FAILED. This in turn sets the status of the operation as a whole to FAILED, and CloudFormation cancels the operation in any remaining Regions.

QUEUED: [Service-managed permissions] For automatic deployments that require a sequence of operations, the operation is queued to be performed. For more information, see the stack set operation status codes in the CloudFormation User Guide.

RUNNING: The operation is currently being performed.

STOPPED: The user has cancelled the operation.

STOPPING: The operation is in the process of stopping, at user request.

SUCCEEDED: The operation completed creating or updating all the specified stacks without exceeding the failure tolerance for the operation.

stackSetOperationSummary_endTimestamp :: Lens' StackSetOperationSummary (Maybe UTCTime)

The time at which the stack set operation ended, across all accounts and Regions specified. Note that this doesn't necessarily mean that the stack set operation was successful, or even attempted, in each account or Region.

stackSetOperationSummary_operationId :: Lens' StackSetOperationSummary (Maybe Text)

The unique ID of the stack set operation.

stackSetOperationSummary_action :: Lens' StackSetOperationSummary (Maybe StackSetOperationAction)

The type of operation: CREATE, UPDATE, or DELETE. Create and delete operations affect only the specified stack instances that are associated with the specified stack set. Update operations affect both the stack set itself as well as all associated stack set instances.