Network.AWS.CloudFormation.Types.StackSetSummary

Description

Documentation

data StackSetSummary

The structures that contain summary information about the specified stack set.

See: newStackSetSummary smart constructor.

Constructors

StackSetSummary'

Fields

status :: Maybe StackSetStatus

The status of the stack set.

permissionModel :: Maybe PermissionModels

Describes how the IAM roles required for stack set operations are created.

With self-managed permissions, you must create the administrator and execution roles required to deploy to target accounts. For more information, see Grant Self-Managed Stack Set Permissions.

With service-managed permissions, StackSets automatically creates the IAM roles required to deploy to accounts managed by Organizations. For more information, see Grant Service-Managed Stack Set Permissions.

lastDriftCheckTimestamp :: Maybe ISO8601

Most recent time when CloudFormation performed a drift detection operation on the stack set. This value will be NULL for any stack set on which drift detection has not yet been performed.

stackSetId :: Maybe Text

The ID of the stack set.

driftStatus :: Maybe StackDriftStatus

Status of the stack set's actual configuration compared to its expected template and parameter configuration. A stack set is considered to have drifted if one or more of its stack instances have drifted from their expected template and parameter configuration.

DRIFTED: One or more of the stack instances belonging to the stack set stack differs from the expected template and parameter configuration. A stack instance is considered to have drifted if one or more of the resources in the associated stack have drifted.

NOT_CHECKED: CloudFormation has not checked the stack set for drift.

IN_SYNC: All of the stack instances belonging to the stack set stack match from the expected template and parameter configuration.

UNKNOWN: This value is reserved for future use.

autoDeployment :: Maybe AutoDeployment

Service-managed permissions

Describes whether StackSets automatically deploys to Organizations accounts that are added to a target organizational unit (OU).

description :: Maybe Text

A description of the stack set that you specify when the stack set is created or updated.

stackSetName :: Maybe Text

The name of the stack set.

newStackSetSummary :: StackSetSummary

Create a value of StackSetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StackSetSummary', stackSetSummary_status - The status of the stack set.

$sel:permissionModel:StackSetSummary', stackSetSummary_permissionModel - Describes how the IAM roles required for stack set operations are created.

With self-managed permissions, you must create the administrator and execution roles required to deploy to target accounts. For more information, see Grant Self-Managed Stack Set Permissions.

With service-managed permissions, StackSets automatically creates the IAM roles required to deploy to accounts managed by Organizations. For more information, see Grant Service-Managed Stack Set Permissions.

$sel:lastDriftCheckTimestamp:StackSetSummary', stackSetSummary_lastDriftCheckTimestamp - Most recent time when CloudFormation performed a drift detection operation on the stack set. This value will be NULL for any stack set on which drift detection has not yet been performed.

$sel:stackSetId:StackSetSummary', stackSetSummary_stackSetId - The ID of the stack set.

$sel:driftStatus:StackSetSummary', stackSetSummary_driftStatus - Status of the stack set's actual configuration compared to its expected template and parameter configuration. A stack set is considered to have drifted if one or more of its stack instances have drifted from their expected template and parameter configuration.

DRIFTED: One or more of the stack instances belonging to the stack set stack differs from the expected template and parameter configuration. A stack instance is considered to have drifted if one or more of the resources in the associated stack have drifted.

NOT_CHECKED: CloudFormation has not checked the stack set for drift.

IN_SYNC: All of the stack instances belonging to the stack set stack match from the expected template and parameter configuration.

UNKNOWN: This value is reserved for future use.

$sel:autoDeployment:StackSetSummary', stackSetSummary_autoDeployment - [Service-managed permissions] Describes whether StackSets automatically deploys to Organizations accounts that are added to a target organizational unit (OU).

$sel:description:StackSetSummary', stackSetSummary_description - A description of the stack set that you specify when the stack set is created or updated.

$sel:stackSetName:StackSetSummary', stackSetSummary_stackSetName - The name of the stack set.

stackSetSummary_status :: Lens' StackSetSummary (Maybe StackSetStatus)

The status of the stack set.

stackSetSummary_permissionModel :: Lens' StackSetSummary (Maybe PermissionModels)

Describes how the IAM roles required for stack set operations are created.

With self-managed permissions, you must create the administrator and execution roles required to deploy to target accounts. For more information, see Grant Self-Managed Stack Set Permissions.

With service-managed permissions, StackSets automatically creates the IAM roles required to deploy to accounts managed by Organizations. For more information, see Grant Service-Managed Stack Set Permissions.

stackSetSummary_lastDriftCheckTimestamp :: Lens' StackSetSummary (Maybe UTCTime)

Most recent time when CloudFormation performed a drift detection operation on the stack set. This value will be NULL for any stack set on which drift detection has not yet been performed.

stackSetSummary_stackSetId :: Lens' StackSetSummary (Maybe Text)

The ID of the stack set.

stackSetSummary_driftStatus :: Lens' StackSetSummary (Maybe StackDriftStatus)

Status of the stack set's actual configuration compared to its expected template and parameter configuration. A stack set is considered to have drifted if one or more of its stack instances have drifted from their expected template and parameter configuration.

DRIFTED: One or more of the stack instances belonging to the stack set stack differs from the expected template and parameter configuration. A stack instance is considered to have drifted if one or more of the resources in the associated stack have drifted.

NOT_CHECKED: CloudFormation has not checked the stack set for drift.

IN_SYNC: All of the stack instances belonging to the stack set stack match from the expected template and parameter configuration.

UNKNOWN: This value is reserved for future use.

stackSetSummary_autoDeployment :: Lens' StackSetSummary (Maybe AutoDeployment)

Service-managed permissions

Describes whether StackSets automatically deploys to Organizations accounts that are added to a target organizational unit (OU).

stackSetSummary_description :: Lens' StackSetSummary (Maybe Text)

A description of the stack set that you specify when the stack set is created or updated.

stackSetSummary_stackSetName :: Lens' StackSetSummary (Maybe Text)

The name of the stack set.