Network.AWS.CloudFormation.Types.TypeConfigurationDetails

Description

Documentation

data TypeConfigurationDetails

Detailed information concerning the specification of a CloudFormation extension in a given account and region.

For more information, see Configuring extensions at the account level in the CloudFormation User Guide.

See: newTypeConfigurationDetails smart constructor.

Constructors

TypeConfigurationDetails'

Fields

lastUpdated :: Maybe ISO8601

When the configuration data was last updated for this extension.

If a configuration has not been set for a specified extension, CloudFormation returns null.

typeName :: Maybe Text

The name of the extension.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the configuration data, in this account and region.

alias :: Maybe Text

The alias specified for this configuration, if one was specified when the configuration was set.

typeArn :: Maybe Text

The Amazon Resource Name (ARN) for the extension, in this account and region.

For public extensions, this will be the ARN assigned when you activate the type in this account and region. For private extensions, this will be the ARN assigned when you register the type in this account and region.

configuration :: Maybe Text

A JSON string specifying the configuration data for the extension, in this account and region.

If a configuration has not been set for a specified extension, CloudFormation returns {}.

isDefaultConfiguration :: Maybe Bool

Whether or not this configuration data is the default configuration for the extension.

newTypeConfigurationDetails :: TypeConfigurationDetails

Create a value of TypeConfigurationDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdated:TypeConfigurationDetails', typeConfigurationDetails_lastUpdated - When the configuration data was last updated for this extension.

If a configuration has not been set for a specified extension, CloudFormation returns null.

$sel:typeName:TypeConfigurationDetails', typeConfigurationDetails_typeName - The name of the extension.

$sel:arn:TypeConfigurationDetails', typeConfigurationDetails_arn - The Amazon Resource Name (ARN) for the configuration data, in this account and region.

$sel:alias:TypeConfigurationDetails', typeConfigurationDetails_alias - The alias specified for this configuration, if one was specified when the configuration was set.

$sel:typeArn:TypeConfigurationDetails', typeConfigurationDetails_typeArn - The Amazon Resource Name (ARN) for the extension, in this account and region.

For public extensions, this will be the ARN assigned when you activate the type in this account and region. For private extensions, this will be the ARN assigned when you register the type in this account and region.

$sel:configuration:TypeConfigurationDetails', typeConfigurationDetails_configuration - A JSON string specifying the configuration data for the extension, in this account and region.

If a configuration has not been set for a specified extension, CloudFormation returns {}.

$sel:isDefaultConfiguration:TypeConfigurationDetails', typeConfigurationDetails_isDefaultConfiguration - Whether or not this configuration data is the default configuration for the extension.

typeConfigurationDetails_lastUpdated :: Lens' TypeConfigurationDetails (Maybe UTCTime)

When the configuration data was last updated for this extension.

If a configuration has not been set for a specified extension, CloudFormation returns null.

typeConfigurationDetails_typeName :: Lens' TypeConfigurationDetails (Maybe Text)

The name of the extension.

typeConfigurationDetails_arn :: Lens' TypeConfigurationDetails (Maybe Text)

The Amazon Resource Name (ARN) for the configuration data, in this account and region.

typeConfigurationDetails_alias :: Lens' TypeConfigurationDetails (Maybe Text)

The alias specified for this configuration, if one was specified when the configuration was set.

typeConfigurationDetails_typeArn :: Lens' TypeConfigurationDetails (Maybe Text)

The Amazon Resource Name (ARN) for the extension, in this account and region.

For public extensions, this will be the ARN assigned when you activate the type in this account and region. For private extensions, this will be the ARN assigned when you register the type in this account and region.

typeConfigurationDetails_configuration :: Lens' TypeConfigurationDetails (Maybe Text)

A JSON string specifying the configuration data for the extension, in this account and region.

If a configuration has not been set for a specified extension, CloudFormation returns {}.

typeConfigurationDetails_isDefaultConfiguration :: Lens' TypeConfigurationDetails (Maybe Bool)

Whether or not this configuration data is the default configuration for the extension.