Network.AWS.CloudFormation.Types.TypeConfigurationIdentifier

Description

Documentation

data TypeConfigurationIdentifier

Identifying information for the configuration of a CloudFormation extension.

See: newTypeConfigurationIdentifier smart constructor.

Constructors

TypeConfigurationIdentifier'

Fields

typeName :: Maybe Text

The name of the extension type to which this configuration applies.

typeConfigurationAlias :: Maybe Text

The alias specified for this configuration, if one was specified when the configuration was set.

type' :: Maybe ThirdPartyType

The type of extension.

typeConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) for the configuration, in this account and region.

typeArn :: Maybe Text

The Amazon Resource Name (ARN) for the extension, in this account and region.

For public extensions, this will be the ARN assigned when you activate the type in this account and region. For private extensions, this will be the ARN assigned when you register the type in this account and region.

newTypeConfigurationIdentifier :: TypeConfigurationIdentifier

Create a value of TypeConfigurationIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:TypeConfigurationIdentifier', typeConfigurationIdentifier_typeName - The name of the extension type to which this configuration applies.

$sel:typeConfigurationAlias:TypeConfigurationIdentifier', typeConfigurationIdentifier_typeConfigurationAlias - The alias specified for this configuration, if one was specified when the configuration was set.

$sel:type':TypeConfigurationIdentifier', typeConfigurationIdentifier_type - The type of extension.

$sel:typeConfigurationArn:TypeConfigurationIdentifier', typeConfigurationIdentifier_typeConfigurationArn - The Amazon Resource Name (ARN) for the configuration, in this account and region.

$sel:typeArn:TypeConfigurationIdentifier', typeConfigurationIdentifier_typeArn - The Amazon Resource Name (ARN) for the extension, in this account and region.

For public extensions, this will be the ARN assigned when you activate the type in this account and region. For private extensions, this will be the ARN assigned when you register the type in this account and region.

typeConfigurationIdentifier_typeName :: Lens' TypeConfigurationIdentifier (Maybe Text)

The name of the extension type to which this configuration applies.

typeConfigurationIdentifier_typeConfigurationAlias :: Lens' TypeConfigurationIdentifier (Maybe Text)

The alias specified for this configuration, if one was specified when the configuration was set.

typeConfigurationIdentifier_type :: Lens' TypeConfigurationIdentifier (Maybe ThirdPartyType)

The type of extension.

typeConfigurationIdentifier_typeConfigurationArn :: Lens' TypeConfigurationIdentifier (Maybe Text)

The Amazon Resource Name (ARN) for the configuration, in this account and region.

typeConfigurationIdentifier_typeArn :: Lens' TypeConfigurationIdentifier (Maybe Text)

The Amazon Resource Name (ARN) for the extension, in this account and region.

For public extensions, this will be the ARN assigned when you activate the type in this account and region. For private extensions, this will be the ARN assigned when you register the type in this account and region.