Network.AWS.CloudFormation.Types.TypeFilters

Description

Documentation

data TypeFilters

Filter criteria to use in determining which extensions to return.

See: newTypeFilters smart constructor.

Constructors

TypeFilters'

Fields

typeNamePrefix :: Maybe Text

A prefix to use as a filter for results.

publisherId :: Maybe Text

The id of the publisher of the extension.

Extensions published by Amazon are not assigned a publisher ID. Use the AWS-TYPES category to specify a list of types published by Amazon.

category :: Maybe Category

The category of extensions to return.

REGISTERED: Private extensions that have been registered for this account and region.

ACTIVATED: Public extensions that have been activated for this account and region.

THIRD-PARTY: Extensions available for use from publishers other than Amazon. This includes:

Private extensions registered in the account.

Public extensions from publishers other than Amazon, whether activated or not.

AWS-TYPES: Extensions available for use from Amazon.

newTypeFilters :: TypeFilters

Create a value of TypeFilters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeNamePrefix:TypeFilters', typeFilters_typeNamePrefix - A prefix to use as a filter for results.

$sel:publisherId:TypeFilters', typeFilters_publisherId - The id of the publisher of the extension.

Extensions published by Amazon are not assigned a publisher ID. Use the AWS-TYPES category to specify a list of types published by Amazon.

$sel:category:TypeFilters', typeFilters_category - The category of extensions to return.

REGISTERED: Private extensions that have been registered for this account and region.

ACTIVATED: Public extensions that have been activated for this account and region.

THIRD-PARTY: Extensions available for use from publishers other than Amazon. This includes:

Private extensions registered in the account.

Public extensions from publishers other than Amazon, whether activated or not.

AWS-TYPES: Extensions available for use from Amazon.

typeFilters_typeNamePrefix :: Lens' TypeFilters (Maybe Text)

A prefix to use as a filter for results.

typeFilters_publisherId :: Lens' TypeFilters (Maybe Text)

The id of the publisher of the extension.

Extensions published by Amazon are not assigned a publisher ID. Use the AWS-TYPES category to specify a list of types published by Amazon.

typeFilters_category :: Lens' TypeFilters (Maybe Category)

The category of extensions to return.

REGISTERED: Private extensions that have been registered for this account and region.

ACTIVATED: Public extensions that have been activated for this account and region.

THIRD-PARTY: Extensions available for use from publishers other than Amazon. This includes:

Private extensions registered in the account.

Public extensions from publishers other than Amazon, whether activated or not.

AWS-TYPES: Extensions available for use from Amazon.