Network.AWS.CloudFormation.UpdateStackInstances

Description

Updates the parameter values for stack instances for the specified accounts, within the specified Regions. A stack instance refers to a stack in a specific account and Region.

You can only update stack instances in Regions and accounts where they already exist; to create additional stack instances, use CreateStackInstances.

During stack set updates, any parameters overridden for a stack instance are not updated, but retain their overridden value.

You can only update the parameter values that are specified in the stack set; to add or delete a parameter itself, use UpdateStackSet to update the stack set template. If you add a parameter to a template, before you can override the parameter value specified in the stack set you must first use UpdateStackSet to update all stack instances with the updated template and parameter value specified in the stack set. Once a stack instance has been updated with the new parameter, you can then override the parameter value using UpdateStackInstances.

Creating a Request

data UpdateStackInstances

See: newUpdateStackInstances smart constructor.

Constructors

UpdateStackInstances'

Fields

parameterOverrides :: Maybe [Parameter]

A list of input parameters whose values you want to update for the specified stack instances.

Any overridden parameter values will be applied to all stack instances in the specified accounts and Regions. When specifying parameters and their values, be aware of how CloudFormation sets parameter values during stack instance update operations:

To override the current value for a parameter, include the parameter and specify its value.

To leave an overridden parameter set to its present value, include the parameter and specify UsePreviousValue as true. (You cannot specify both a value and set UsePreviousValue to true.)

To set an overridden parameter back to the value specified in the stack set, specify a parameter list but do not include the parameter in the list.

To leave all parameters set to their present values, do not specify this property at all.

During stack set updates, any parameter values overridden for a stack instance are not updated, but retain their overridden value.

You can only override the parameter values that are specified in the stack set; to add or delete a parameter itself, use UpdateStackSet to update the stack set template. If you add a parameter to a template, before you can override the parameter value specified in the stack set you must first use UpdateStackSet to update all stack instances with the updated template and parameter value specified in the stack set. Once a stack instance has been updated with the new parameter, you can then override the parameter value using UpdateStackInstances.

deploymentTargets :: Maybe DeploymentTargets

Service-managed permissions

The Organizations accounts for which you want to update parameter values for stack instances. If your update targets OUs, the overridden parameter values only apply to the accounts that are currently in the target OUs and their child OUs. Accounts added to the target OUs and their child OUs in the future won't use the overridden values.

You can specify Accounts or DeploymentTargets, but not both.

operationId :: Maybe Text

The unique identifier for this stack set operation.

The operation ID also functions as an idempotency token, to ensure that CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You might retry stack set operation requests to ensure that CloudFormation successfully received them.

If you don't specify an operation ID, the SDK generates one automatically.

callAs :: Maybe CallAs

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

operationPreferences :: Maybe StackSetOperationPreferences

Preferences for how CloudFormation performs this stack set operation.

accounts :: Maybe [Text]

Self-managed permissions

The names of one or more Amazon Web Services accounts for which you want to update parameter values for stack instances. The overridden parameter values will be applied to all stack instances in the specified accounts and Regions.

You can specify Accounts or DeploymentTargets, but not both.

stackSetName :: Text

The name or unique ID of the stack set associated with the stack instances.

regions :: [Text]

The names of one or more Regions in which you want to update parameter values for stack instances. The overridden parameter values will be applied to all stack instances in the specified accounts and Regions.

newUpdateStackInstances

Arguments

:: Text
$sel:stackSetName:UpdateStackInstances'

-> UpdateStackInstances

Create a value of UpdateStackInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameterOverrides:UpdateStackInstances', updateStackInstances_parameterOverrides - A list of input parameters whose values you want to update for the specified stack instances.

Any overridden parameter values will be applied to all stack instances in the specified accounts and Regions. When specifying parameters and their values, be aware of how CloudFormation sets parameter values during stack instance update operations:

To override the current value for a parameter, include the parameter and specify its value.

To leave an overridden parameter set to its present value, include the parameter and specify UsePreviousValue as true. (You cannot specify both a value and set UsePreviousValue to true.)

To set an overridden parameter back to the value specified in the stack set, specify a parameter list but do not include the parameter in the list.

To leave all parameters set to their present values, do not specify this property at all.

During stack set updates, any parameter values overridden for a stack instance are not updated, but retain their overridden value.

You can only override the parameter values that are specified in the stack set; to add or delete a parameter itself, use UpdateStackSet to update the stack set template. If you add a parameter to a template, before you can override the parameter value specified in the stack set you must first use UpdateStackSet to update all stack instances with the updated template and parameter value specified in the stack set. Once a stack instance has been updated with the new parameter, you can then override the parameter value using UpdateStackInstances.

$sel:deploymentTargets:UpdateStackInstances', updateStackInstances_deploymentTargets - [Service-managed permissions] The Organizations accounts for which you want to update parameter values for stack instances. If your update targets OUs, the overridden parameter values only apply to the accounts that are currently in the target OUs and their child OUs. Accounts added to the target OUs and their child OUs in the future won't use the overridden values.

You can specify Accounts or DeploymentTargets, but not both.

$sel:operationId:UpdateStackInstances', updateStackInstances_operationId - The unique identifier for this stack set operation.

The operation ID also functions as an idempotency token, to ensure that CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You might retry stack set operation requests to ensure that CloudFormation successfully received them.

If you don't specify an operation ID, the SDK generates one automatically.

$sel:callAs:UpdateStackInstances', updateStackInstances_callAs - [Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

$sel:operationPreferences:UpdateStackInstances', updateStackInstances_operationPreferences - Preferences for how CloudFormation performs this stack set operation.

$sel:accounts:UpdateStackInstances', updateStackInstances_accounts - [Self-managed permissions] The names of one or more Amazon Web Services accounts for which you want to update parameter values for stack instances. The overridden parameter values will be applied to all stack instances in the specified accounts and Regions.

You can specify Accounts or DeploymentTargets, but not both.

$sel:stackSetName:UpdateStackInstances', updateStackInstances_stackSetName - The name or unique ID of the stack set associated with the stack instances.

$sel:regions:UpdateStackInstances', updateStackInstances_regions - The names of one or more Regions in which you want to update parameter values for stack instances. The overridden parameter values will be applied to all stack instances in the specified accounts and Regions.

Request Lenses

updateStackInstances_parameterOverrides :: Lens' UpdateStackInstances (Maybe [Parameter])

A list of input parameters whose values you want to update for the specified stack instances.

Any overridden parameter values will be applied to all stack instances in the specified accounts and Regions. When specifying parameters and their values, be aware of how CloudFormation sets parameter values during stack instance update operations:

To override the current value for a parameter, include the parameter and specify its value.

To leave an overridden parameter set to its present value, include the parameter and specify UsePreviousValue as true. (You cannot specify both a value and set UsePreviousValue to true.)

To set an overridden parameter back to the value specified in the stack set, specify a parameter list but do not include the parameter in the list.

To leave all parameters set to their present values, do not specify this property at all.

During stack set updates, any parameter values overridden for a stack instance are not updated, but retain their overridden value.

You can only override the parameter values that are specified in the stack set; to add or delete a parameter itself, use UpdateStackSet to update the stack set template. If you add a parameter to a template, before you can override the parameter value specified in the stack set you must first use UpdateStackSet to update all stack instances with the updated template and parameter value specified in the stack set. Once a stack instance has been updated with the new parameter, you can then override the parameter value using UpdateStackInstances.

updateStackInstances_deploymentTargets :: Lens' UpdateStackInstances (Maybe DeploymentTargets)

Service-managed permissions

The Organizations accounts for which you want to update parameter values for stack instances. If your update targets OUs, the overridden parameter values only apply to the accounts that are currently in the target OUs and their child OUs. Accounts added to the target OUs and their child OUs in the future won't use the overridden values.

You can specify Accounts or DeploymentTargets, but not both.

updateStackInstances_operationId :: Lens' UpdateStackInstances (Maybe Text)

The unique identifier for this stack set operation.

The operation ID also functions as an idempotency token, to ensure that CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You might retry stack set operation requests to ensure that CloudFormation successfully received them.

If you don't specify an operation ID, the SDK generates one automatically.

updateStackInstances_callAs :: Lens' UpdateStackInstances (Maybe CallAs)

Service-managed permissions

Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your Amazon Web Services account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the CloudFormation User Guide.

updateStackInstances_operationPreferences :: Lens' UpdateStackInstances (Maybe StackSetOperationPreferences)

Preferences for how CloudFormation performs this stack set operation.

updateStackInstances_accounts :: Lens' UpdateStackInstances (Maybe [Text])

Self-managed permissions

The names of one or more Amazon Web Services accounts for which you want to update parameter values for stack instances. The overridden parameter values will be applied to all stack instances in the specified accounts and Regions.

You can specify Accounts or DeploymentTargets, but not both.

updateStackInstances_stackSetName :: Lens' UpdateStackInstances Text

The name or unique ID of the stack set associated with the stack instances.

updateStackInstances_regions :: Lens' UpdateStackInstances [Text]

The names of one or more Regions in which you want to update parameter values for stack instances. The overridden parameter values will be applied to all stack instances in the specified accounts and Regions.

Destructuring the Response

data UpdateStackInstancesResponse

See: newUpdateStackInstancesResponse smart constructor.

Constructors

UpdateStackInstancesResponse'

Fields

operationId :: Maybe Text

The unique identifier for this stack set operation.

httpStatus :: Int

The response's http status code.

newUpdateStackInstancesResponse

Arguments

:: Int
$sel:httpStatus:UpdateStackInstancesResponse'

-> UpdateStackInstancesResponse

Create a value of UpdateStackInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationId:UpdateStackInstances', updateStackInstancesResponse_operationId - The unique identifier for this stack set operation.

$sel:httpStatus:UpdateStackInstancesResponse', updateStackInstancesResponse_httpStatus - The response's http status code.

Response Lenses

updateStackInstancesResponse_operationId :: Lens' UpdateStackInstancesResponse (Maybe Text)

The unique identifier for this stack set operation.

updateStackInstancesResponse_httpStatus :: Lens' UpdateStackInstancesResponse Int

The response's http status code.