Network.AWS.CloudFormation.V2010_05_15.CancelUpdateStack

Description

Cancels an update on the specified stack. If the call completes successfully, the stack will roll back the update and revert to the previous stack configuration. Only stacks that are in the UPDATE_IN_PROGRESS state can be canceled. https:/cloudformation.us-east-1.amazonaws.com ?Action=CancelUpdateStack &StackName=MyStack &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2010-07-27T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature].

Documentation

data CancelUpdateStack

Constructors

CancelUpdateStack

Fields

_cusiStackName :: Text

The name or the unique identifier associated with the stack.

cusiStackName :: Iso' CancelUpdateStack Text

data CancelUpdateStackResponse

Constructors

CancelUpdateStackResponse