Network.AWS.CloudFormation.V2010_05_15.DescribeStacks

Description

Returns the description for the specified stack; if no stack name was specified, then it returns the description for all the stacks created. https:/cloudformation.us-east-1.amazonaws.com ?Action=DescribeStacks &StackName=MyStack &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2010-07-27T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature] MyStack arn:aws:cloudformation:us-east-1:123456789:stackMyStackaaf549a0-a413-11df-adb3-5081b3858e83 2010-07-27T22:28:28Z CREATE_COMPLETE false StartPage http:/my-load-balancer.amazonaws.com:80index.html.

Documentation

describeStacks :: DescribeStacks

Minimum specification for a DescribeStacks request.

data DescribeStacks

Constructors

DescribeStacks

Fields

_dsjNextToken :: Maybe Text

String that identifies the start of the next list of stacks, if there is one.

_dsjStackName :: Maybe Text

The name or the unique identifier associated with the stack, which are not always interchangeable: Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

dsjStackName :: Lens' DescribeStacks (Maybe Text)

dsjNextToken :: Lens' DescribeStacks (Maybe Text)

data DescribeStacksResponse

Constructors

DescribeStacksResponse

Fields

_dsoNextToken :: Maybe Text

String that identifies the start of the next list of stacks, if there is one.

_dsoStacks :: [Stack]

A list of stack structures.

dsoStacks :: Lens' DescribeStacksResponse [Stack]

dsoNextToken :: Lens' DescribeStacksResponse (Maybe Text)