Network.AWS.CloudFormation.V2010_05_15.GetStackPolicy

Description

Returns the stack policy for a specified stack. If a stack doesn't have a policy, a null value is returned. https:/cloudformation.us-east-1.amazonaws.com ?Action=GetStackPolicy &StackName=MyStack &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2010-07-27T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature] "{ Statement : [ { Effect : Deny, Action : "Update:*", Principal : "*", Resource : "LogicalResourceId/ProductionDatabase" }, { Effect : Allow, Action : "Update:*", Principal : "*", Resource : "*" } ] }.

Documentation

data GetStackPolicy

Constructors

GetStackPolicy

Fields

_gspiStackName :: Text

The name or stack ID that is associated with the stack whose policy you want to get.

gspiStackName :: Iso' GetStackPolicy Text

data GetStackPolicyResponse

Constructors

GetStackPolicyResponse

Fields

_gspoStackPolicyBody :: Maybe Text

Structure containing the stack policy body. (For more information, go to Prevent Updates to Stack Resources in the AWS CloudFormation User Guide.).

gspoStackPolicyBody :: Iso' GetStackPolicyResponse (Maybe Text)