Network.AWS.CloudFormation.V2010_05_15.GetTemplate

Description

Returns the template body for a specified stack. You can get the template for running or deleted stacks. For deleted stacks, GetTemplate returns the template for up to 90 days after the stack has been deleted. If the template does not exist, a ValidationError is returned. https:/cloudformation.us-east-1.amazonaws.com ?Action=GetTemplate &StackName=MyStack &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2010-07-27T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature] "{ AWSTemplateFormatVersion : "2010-09-09", Description : "Simple example", Resources : { MySQS : { Type : AWS::SQS::Queue, Properties : { } } } }.

Documentation

data GetTemplate

Constructors

GetTemplate

Fields

_gtiStackName :: Text

The name or the unique identifier associated with the stack, which are not always interchangeable: Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

gtiStackName :: Iso' GetTemplate Text

data GetTemplateResponse

Constructors

GetTemplateResponse

Fields

_gtoTemplateBody :: Maybe Text

Structure containing the template body. (For more information, go to Template Anatomy in the AWS CloudFormation User Guide.).

gtoTemplateBody :: Iso' GetTemplateResponse (Maybe Text)