Network.AWS.CloudFormation.V2010_05_15.SetStackPolicy

Description

Sets a stack policy for a specified stack. https:/cloudformation.us-east-1.amazonaws.com ?Action=SetStackPolicy &StackName=MyStack &StackPolicyBody=[Stack Policy Document] &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2010-07-27T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature].

Documentation

setStackPolicy

Arguments

:: Text
_sspiStackName

-> SetStackPolicy

Minimum specification for a SetStackPolicy request.

data SetStackPolicy

Constructors

SetStackPolicy

Fields

_sspiStackName :: Text

The name or stack ID that you want to associate a policy with.

_sspiStackPolicyBody :: Maybe Text

Structure containing the stack policy body. For more information, go to Prevent Updates to Stack Resources in the AWS CloudFormation User Guide. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

_sspiStackPolicyURL :: Maybe Text

Location of a file containing the stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

sspiStackPolicyURL :: Lens' SetStackPolicy (Maybe Text)

sspiStackPolicyBody :: Lens' SetStackPolicy (Maybe Text)

sspiStackName :: Lens' SetStackPolicy Text

data SetStackPolicyResponse

Constructors

SetStackPolicyResponse