Network.AWS.CloudFormation.V2010_05_15.Types

Description

AWS CloudFormation gives developers and systems administrators an easy way to create and manage a collection of related AWS resources, provisioning and updating them in an orderly and predictable fashion. You can use AWS CloudFormation’s sample templates or create your own templates to describe the AWS resources, and any associated dependencies or runtime parameters, required to run your application. You don’t need to figure out the order for provisioning AWS services or the subtleties of making those dependencies work. CloudFormation takes care of this for you. After the AWS resources are deployed, you can modify and update them in a controlled and predictable way, in effect applying version control to your AWS infrastructure the same way you do with your software.

Documentation

data CloudFormation

Supported version (2010-05-15) of the AWS CloudFormation service.

xmlOptions :: Tagged a XMLOptions

data Capability

Constructors

CapabilityCapabilityIam
CAPABILITY_IAM

data OnFailure

Determines what action will be taken if stack creation fails. This must be one of: DO_NOTHING, ROLLBACK, or DELETE. You can specify either OnFailure or DisableRollback, but not both. Default: ROLLBACK.

Constructors

OnFailureDelete
DELETE

OnFailureDoNothing
DO_NOTHING

OnFailureRollback
ROLLBACK

data ResourceStatus

Current status of the resource.

Constructors

ResourceStatusCreateComplete
CREATE_COMPLETE

ResourceStatusCreateFailed
CREATE_FAILED

ResourceStatusCreateInProgress
CREATE_IN_PROGRESS

ResourceStatusDeleteComplete
DELETE_COMPLETE

ResourceStatusDeleteFailed
DELETE_FAILED

ResourceStatusDeleteInProgress
DELETE_IN_PROGRESS

ResourceStatusUpdateComplete
UPDATE_COMPLETE

ResourceStatusUpdateFailed
UPDATE_FAILED

ResourceStatusUpdateInProgress
UPDATE_IN_PROGRESS

data StackStatus

Current status of the stack.

Constructors

StackStatusCreateComplete
CREATE_COMPLETE

StackStatusCreateFailed
CREATE_FAILED

StackStatusCreateInProgress
CREATE_IN_PROGRESS

StackStatusDeleteComplete
DELETE_COMPLETE

StackStatusDeleteFailed
DELETE_FAILED

StackStatusDeleteInProgress
DELETE_IN_PROGRESS

StackStatusRollbackComplete
ROLLBACK_COMPLETE

StackStatusRollbackFailed
ROLLBACK_FAILED

StackStatusRollbackInProgress
ROLLBACK_IN_PROGRESS

StackStatusUpdateComplete
UPDATE_COMPLETE

StackStatusUpdateCompleteCleanupInProgress
UPDATE_COMPLETE_CLEANUP_IN_PROGRESS

StackStatusUpdateInProgress
UPDATE_IN_PROGRESS

StackStatusUpdateRollbackComplete
UPDATE_ROLLBACK_COMPLETE

StackStatusUpdateRollbackCompleteCleanupInProgress
UPDATE_ROLLBACK_COMPLETE_CLEANUP_IN_PROGRESS

StackStatusUpdateRollbackFailed
UPDATE_ROLLBACK_FAILED

StackStatusUpdateRollbackInProgress
UPDATE_ROLLBACK_IN_PROGRESS

data Output

The Output data type.

Constructors

Output

Fields

_rOutputValue :: Maybe Text

The value associated with the output.

_rOutputKey :: Maybe Text

The key associated with the output.

_rDescription :: Maybe Text

User defined description associated with the output.

data Parameter

The Parameter data type.

Constructors

Parameter

Fields

_qParameterValue :: Maybe Text

The value associated with the parameter.

_qParameterKey :: Maybe Text

The key associated with the parameter.

_qUsePreviousValue :: Maybe Bool

During a stack update, use the existing parameter value that is being used for the stack.

data Stack

The Stack data type.

Constructors

Stack

Fields

_uCreationTime :: ISO8601

Time at which the stack was created.

_uStackStatus :: StackStatus

Current status of the stack.

_uDisableRollback :: Maybe Bool

Boolean to enable or disable rollback on stack creation failures: true: disable rollback false: enable rollback.

_uLastUpdatedTime :: Maybe ISO8601

The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

_uNotificationARNs :: [Text]

SNS topic ARNs to which stack related events are published.

_uStackStatusReason :: Maybe Text

Success/failure message associated with the stack status.

_uOutputs :: [Output]

A list of output structures.

_uParameters :: [Parameter]

A list of Parameter structures.

_uStackId :: Maybe Text

Unique identifier of the stack.

_uDescription :: Maybe Text

User defined description associated with the stack.

_uCapabilities :: [Capability]

The capabilities allowed in the stack.

_uTags :: [Tag]

A list of Tags that specify cost allocation information for the stack.

_uTimeoutInMinutes :: Maybe Integer

The amount of time within which stack creation should complete.

_uStackName :: Text

The name associated with the stack.

data StackEvent

The StackEvent data type.

Constructors

StackEvent

Fields

_sfLogicalResourceId :: Maybe Text

The logical name of the resource specified in the template.

_sfPhysicalResourceId :: Maybe Text

The name or unique identifier associated with the physical instance of the resource.

_sfResourceType :: Maybe Text

Type of resource. (For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.).

_sfResourceStatusReason :: Maybe Text

Success/failure message associated with the resource.

_sfResourceProperties :: Maybe Text

BLOB of the properties used to create the resource.

_sfResourceStatus :: Maybe ResourceStatus

Current status of the resource.

_sfStackId :: Text

The unique ID name of the instance of the stack.

_sfTimestamp :: ISO8601

Time the status was updated.

_sfEventId :: Text

The unique ID of this event.

_sfStackName :: Text

The name associated with a stack.

data StackResource

The StackResource data type.

Constructors

StackResource

Fields

_sshLogicalResourceId :: Text

The logical name of the resource specified in the template.

_sshPhysicalResourceId :: Maybe Text

The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation.

_sshResourceType :: Text

Type of resource. (For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.).

_sshResourceStatusReason :: Maybe Text

Success/failure message associated with the resource.

_sshResourceStatus :: ResourceStatus

Current status of the resource.

_sshStackId :: Maybe Text

Unique identifier of the stack.

_sshDescription :: Maybe Text

User defined description associated with the resource.

_sshTimestamp :: ISO8601

Time the status was updated.

_sshStackName :: Maybe Text

The name associated with the stack.

data StackResourceDetail

A StackResourceDetail structure containing the description of the specified resource in the specified stack.

Constructors

StackResourceDetail

Fields

_srdLogicalResourceId :: Text

The logical name of the resource specified in the template.

_srdPhysicalResourceId :: Maybe Text

The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation.

_srdResourceType :: Text

Type of resource. ((For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.).

_srdResourceStatusReason :: Maybe Text

Success/failure message associated with the resource.

_srdResourceStatus :: ResourceStatus

Current status of the resource.

_srdMetadata :: Maybe Text

The JSON format content of the Metadata attribute declared for the resource. For more information, see Metadata Attribute in the AWS CloudFormation User Guide.

_srdStackId :: Maybe Text

Unique identifier of the stack.

_srdDescription :: Maybe Text

User defined description associated with the resource.

_srdLastUpdatedTimestamp :: ISO8601

Time the status was updated.

_srdStackName :: Maybe Text

The name associated with the stack.

data StackResourceSummary

Contains high-level information about the specified stack resource.

Constructors

StackResourceSummary

Fields

_srtLogicalResourceId :: Text

The logical name of the resource specified in the template.

_srtPhysicalResourceId :: Maybe Text

The name or unique identifier that corresponds to a physical instance ID of the resource.

_srtResourceType :: Text

Type of resource. (For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.).

_srtResourceStatusReason :: Maybe Text

Success/failure message associated with the resource.

_srtResourceStatus :: ResourceStatus

Current status of the resource.

_srtLastUpdatedTimestamp :: ISO8601

Time the status was updated.

data StackSummary

The StackSummary Data Type.

Constructors

StackSummary

Fields

_ssyCreationTime :: ISO8601

The time the stack was created.

_ssyStackStatus :: StackStatus

The current status of the stack.

_ssyLastUpdatedTime :: Maybe ISO8601

The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

_ssyStackStatusReason :: Maybe Text

Success/Failure message associated with the stack status.

_ssyTemplateDescription :: Maybe Text

The template description of the template used to create the stack.

_ssyDeletionTime :: Maybe ISO8601

The time the stack was deleted.

_ssyStackId :: Maybe Text

Unique stack identifier.

_ssyStackName :: Text

The name associated with the stack.

data Tag

The Tag type is used by CreateStack in the Tags parameter. It allows you to specify a key/value pair that can be used to store information related to cost allocation for an AWS CloudFormation stack.

Constructors

Tag

Fields

_yValue :: Maybe Text

Required. A string containing the value for this tag. You can specify a maximum of 256 characters for a tag value.

_yKey :: Maybe Text

Required. A string used to identify this tag. You can specify a maximum of 128 characters for a tag key. Tags owned by Amazon Web Services (AWS) have the reserved prefix: aws:.

data TemplateParameter

The TemplateParameter data type.

Constructors

TemplateParameter

Fields

_tqParameterKey :: Maybe Text

The name associated with the parameter.

_tqDefaultValue :: Maybe Text

The default value associated with the parameter.

_tqNoEcho :: Maybe Bool

Flag indicating whether the parameter should be displayed as plain text in logs and UIs.

_tqDescription :: Maybe Text

User defined description associated with the parameter.

rOutputValue :: Lens' Output (Maybe Text)

rOutputKey :: Lens' Output (Maybe Text)

rDescription :: Lens' Output (Maybe Text)

qUsePreviousValue :: Lens' Parameter (Maybe Bool)

qParameterValue :: Lens' Parameter (Maybe Text)

qParameterKey :: Lens' Parameter (Maybe Text)

uTimeoutInMinutes :: Lens' Stack (Maybe Integer)

uTags :: Lens' Stack [Tag]

uStackStatusReason :: Lens' Stack (Maybe Text)

uStackStatus :: Lens' Stack StackStatus

uStackName :: Lens' Stack Text

uStackId :: Lens' Stack (Maybe Text)

uParameters :: Lens' Stack [Parameter]

uOutputs :: Lens' Stack [Output]

uNotificationARNs :: Lens' Stack [Text]

uLastUpdatedTime :: Lens' Stack (Maybe ISO8601)

uDisableRollback :: Lens' Stack (Maybe Bool)

uDescription :: Lens' Stack (Maybe Text)

uCreationTime :: Lens' Stack ISO8601

uCapabilities :: Lens' Stack [Capability]

sfTimestamp :: Lens' StackEvent ISO8601

sfStackName :: Lens' StackEvent Text

sfStackId :: Lens' StackEvent Text

sfResourceType :: Lens' StackEvent (Maybe Text)

sfResourceStatusReason :: Lens' StackEvent (Maybe Text)

sfResourceStatus :: Lens' StackEvent (Maybe ResourceStatus)

sfResourceProperties :: Lens' StackEvent (Maybe Text)

sfPhysicalResourceId :: Lens' StackEvent (Maybe Text)

sfLogicalResourceId :: Lens' StackEvent (Maybe Text)

sfEventId :: Lens' StackEvent Text

sshTimestamp :: Lens' StackResource ISO8601

sshStackName :: Lens' StackResource (Maybe Text)

sshStackId :: Lens' StackResource (Maybe Text)

sshResourceType :: Lens' StackResource Text

sshResourceStatusReason :: Lens' StackResource (Maybe Text)

sshResourceStatus :: Lens' StackResource ResourceStatus

sshPhysicalResourceId :: Lens' StackResource (Maybe Text)

sshLogicalResourceId :: Lens' StackResource Text

sshDescription :: Lens' StackResource (Maybe Text)

srdStackName :: Lens' StackResourceDetail (Maybe Text)

srdStackId :: Lens' StackResourceDetail (Maybe Text)

srdResourceType :: Lens' StackResourceDetail Text

srdResourceStatusReason :: Lens' StackResourceDetail (Maybe Text)

srdResourceStatus :: Lens' StackResourceDetail ResourceStatus

srdPhysicalResourceId :: Lens' StackResourceDetail (Maybe Text)

srdMetadata :: Lens' StackResourceDetail (Maybe Text)

srdLogicalResourceId :: Lens' StackResourceDetail Text

srdLastUpdatedTimestamp :: Lens' StackResourceDetail ISO8601

srdDescription :: Lens' StackResourceDetail (Maybe Text)

srtResourceType :: Lens' StackResourceSummary Text

srtResourceStatusReason :: Lens' StackResourceSummary (Maybe Text)

srtResourceStatus :: Lens' StackResourceSummary ResourceStatus

srtPhysicalResourceId :: Lens' StackResourceSummary (Maybe Text)

srtLogicalResourceId :: Lens' StackResourceSummary Text

srtLastUpdatedTimestamp :: Lens' StackResourceSummary ISO8601

ssyTemplateDescription :: Lens' StackSummary (Maybe Text)

ssyStackStatusReason :: Lens' StackSummary (Maybe Text)

ssyStackStatus :: Lens' StackSummary StackStatus

ssyStackName :: Lens' StackSummary Text

ssyStackId :: Lens' StackSummary (Maybe Text)

ssyLastUpdatedTime :: Lens' StackSummary (Maybe ISO8601)

ssyDeletionTime :: Lens' StackSummary (Maybe ISO8601)

ssyCreationTime :: Lens' StackSummary ISO8601

yValue :: Lens' Tag (Maybe Text)

yKey :: Lens' Tag (Maybe Text)

tqParameterKey :: Lens' TemplateParameter (Maybe Text)

tqNoEcho :: Lens' TemplateParameter (Maybe Bool)

tqDescription :: Lens' TemplateParameter (Maybe Text)

tqDefaultValue :: Lens' TemplateParameter (Maybe Text)