Network.AWS.CloudFormation.ValidateTemplate

Description

Validates a specified template. CloudFormation first checks if the template is valid JSON. If it isn't, CloudFormation checks if the template is valid YAML. If both these checks fail, CloudFormation returns a template validation error.

Creating a Request

data ValidateTemplate

The input for ValidateTemplate action.

See: newValidateTemplate smart constructor.

Constructors

ValidateTemplate'

Fields

templateBody :: Maybe Text

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, go to Template Anatomy in the CloudFormation User Guide.

Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

templateURL :: Maybe Text

Location of file containing the template body. The URL must point to a template (max size: 460,800 bytes) that is located in an Amazon S3 bucket or a Systems Manager document. For more information, go to Template Anatomy in the CloudFormation User Guide.

Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

newValidateTemplate :: ValidateTemplate

Create a value of ValidateTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateBody:ValidateTemplate', validateTemplate_templateBody - Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, go to Template Anatomy in the CloudFormation User Guide.

Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

$sel:templateURL:ValidateTemplate', validateTemplate_templateURL - Location of file containing the template body. The URL must point to a template (max size: 460,800 bytes) that is located in an Amazon S3 bucket or a Systems Manager document. For more information, go to Template Anatomy in the CloudFormation User Guide.

Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

Request Lenses

validateTemplate_templateBody :: Lens' ValidateTemplate (Maybe Text)

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, go to Template Anatomy in the CloudFormation User Guide.

Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

validateTemplate_templateURL :: Lens' ValidateTemplate (Maybe Text)

Location of file containing the template body. The URL must point to a template (max size: 460,800 bytes) that is located in an Amazon S3 bucket or a Systems Manager document. For more information, go to Template Anatomy in the CloudFormation User Guide.

Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

Destructuring the Response

data ValidateTemplateResponse

The output for ValidateTemplate action.

See: newValidateTemplateResponse smart constructor.

Constructors

ValidateTemplateResponse'

Fields

declaredTransforms :: Maybe [Text]

A list of the transforms that are declared in the template.

capabilitiesReason :: Maybe Text

The list of resources that generated the values in the Capabilities response element.

parameters :: Maybe [TemplateParameter]

A list of TemplateParameter structures.

description :: Maybe Text

The description found within the template.

capabilities :: Maybe [Capability]

The capabilities found within the template. If your template contains IAM resources, you must specify the CAPABILITY_IAM or CAPABILITY_NAMED_IAM value for this parameter when you use the CreateStack or UpdateStack actions with your template; otherwise, those actions return an InsufficientCapabilities error.

For more information, see Acknowledging IAM Resources in CloudFormation Templates.

httpStatus :: Int

The response's http status code.

newValidateTemplateResponse

Arguments

:: Int
$sel:httpStatus:ValidateTemplateResponse'

-> ValidateTemplateResponse

Create a value of ValidateTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:declaredTransforms:ValidateTemplateResponse', validateTemplateResponse_declaredTransforms - A list of the transforms that are declared in the template.

$sel:capabilitiesReason:ValidateTemplateResponse', validateTemplateResponse_capabilitiesReason - The list of resources that generated the values in the Capabilities response element.

$sel:parameters:ValidateTemplateResponse', validateTemplateResponse_parameters - A list of TemplateParameter structures.

$sel:description:ValidateTemplateResponse', validateTemplateResponse_description - The description found within the template.

$sel:capabilities:ValidateTemplateResponse', validateTemplateResponse_capabilities - The capabilities found within the template. If your template contains IAM resources, you must specify the CAPABILITY_IAM or CAPABILITY_NAMED_IAM value for this parameter when you use the CreateStack or UpdateStack actions with your template; otherwise, those actions return an InsufficientCapabilities error.

For more information, see Acknowledging IAM Resources in CloudFormation Templates.

$sel:httpStatus:ValidateTemplateResponse', validateTemplateResponse_httpStatus - The response's http status code.

Response Lenses

validateTemplateResponse_declaredTransforms :: Lens' ValidateTemplateResponse (Maybe [Text])

A list of the transforms that are declared in the template.

validateTemplateResponse_capabilitiesReason :: Lens' ValidateTemplateResponse (Maybe Text)

The list of resources that generated the values in the Capabilities response element.

validateTemplateResponse_parameters :: Lens' ValidateTemplateResponse (Maybe [TemplateParameter])

A list of TemplateParameter structures.

validateTemplateResponse_description :: Lens' ValidateTemplateResponse (Maybe Text)

The description found within the template.

validateTemplateResponse_capabilities :: Lens' ValidateTemplateResponse (Maybe [Capability])

The capabilities found within the template. If your template contains IAM resources, you must specify the CAPABILITY_IAM or CAPABILITY_NAMED_IAM value for this parameter when you use the CreateStack or UpdateStack actions with your template; otherwise, those actions return an InsufficientCapabilities error.

For more information, see Acknowledging IAM Resources in CloudFormation Templates.

validateTemplateResponse_httpStatus :: Lens' ValidateTemplateResponse Int

The response's http status code.