Network.AWS.CloudFront.AssociateAlias

Description

Associates an alias (also known as a CNAME or an alternate domain name) with a CloudFront distribution.

With this operation you can move an alias that’s already in use on a CloudFront distribution to a different distribution in one step. This prevents the downtime that could occur if you first remove the alias from one distribution and then separately add the alias to another distribution.

To use this operation to associate an alias with a distribution, you provide the alias and the ID of the target distribution for the alias. For more information, including how to set up the target distribution, prerequisites that you must complete, and other restrictions, see Moving an alternate domain name to a different distribution in the Amazon CloudFront Developer Guide.

Creating a Request

data AssociateAlias

See: newAssociateAlias smart constructor.

Constructors

AssociateAlias'

Fields

targetDistributionId :: Text

The ID of the distribution that you’re associating the alias with.

alias :: Text

The alias (also known as a CNAME) to add to the target distribution.

newAssociateAlias

Arguments

:: Text
$sel:targetDistributionId:AssociateAlias'

-> Text
$sel:alias:AssociateAlias'

-> AssociateAlias

Create a value of AssociateAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetDistributionId:AssociateAlias', associateAlias_targetDistributionId - The ID of the distribution that you’re associating the alias with.

$sel:alias:AssociateAlias', associateAlias_alias - The alias (also known as a CNAME) to add to the target distribution.

Request Lenses

associateAlias_targetDistributionId :: Lens' AssociateAlias Text

The ID of the distribution that you’re associating the alias with.

associateAlias_alias :: Lens' AssociateAlias Text

The alias (also known as a CNAME) to add to the target distribution.

Destructuring the Response

data AssociateAliasResponse

See: newAssociateAliasResponse smart constructor.

Constructors

AssociateAliasResponse'

newAssociateAliasResponse :: AssociateAliasResponse

Create a value of AssociateAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.