Network.AWS.CloudFront.CreateDistributionWithTags

Description

Create a new distribution with tags.

Creating a Request

data CreateDistributionWithTags

The request to create a new distribution with tags.

See: newCreateDistributionWithTags smart constructor.

Constructors

CreateDistributionWithTags'

Fields

distributionConfigWithTags :: DistributionConfigWithTags

The distribution's configuration information.

newCreateDistributionWithTags

Arguments

:: DistributionConfigWithTags
$sel:distributionConfigWithTags:CreateDistributionWithTags'

-> CreateDistributionWithTags

Create a value of CreateDistributionWithTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:distributionConfigWithTags:CreateDistributionWithTags', createDistributionWithTags_distributionConfigWithTags - The distribution's configuration information.

Request Lenses

createDistributionWithTags_distributionConfigWithTags :: Lens' CreateDistributionWithTags DistributionConfigWithTags

The distribution's configuration information.

Destructuring the Response

data CreateDistributionWithTagsResponse

The returned result of the corresponding request.

See: newCreateDistributionWithTagsResponse smart constructor.

Constructors

CreateDistributionWithTagsResponse'

Fields

eTag :: Maybe Text

The current version of the distribution created.

distribution :: Maybe Distribution

The distribution's information.

location :: Maybe Text

The fully qualified URI of the new distribution resource just created.

httpStatus :: Int

The response's http status code.

newCreateDistributionWithTagsResponse

Arguments

:: Int
$sel:httpStatus:CreateDistributionWithTagsResponse'

-> CreateDistributionWithTagsResponse

Create a value of CreateDistributionWithTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:CreateDistributionWithTagsResponse', createDistributionWithTagsResponse_eTag - The current version of the distribution created.

$sel:distribution:CreateDistributionWithTagsResponse', createDistributionWithTagsResponse_distribution - The distribution's information.

$sel:location:CreateDistributionWithTagsResponse', createDistributionWithTagsResponse_location - The fully qualified URI of the new distribution resource just created.

$sel:httpStatus:CreateDistributionWithTagsResponse', createDistributionWithTagsResponse_httpStatus - The response's http status code.

Response Lenses

createDistributionWithTagsResponse_eTag :: Lens' CreateDistributionWithTagsResponse (Maybe Text)

The current version of the distribution created.

createDistributionWithTagsResponse_distribution :: Lens' CreateDistributionWithTagsResponse (Maybe Distribution)

The distribution's information.

createDistributionWithTagsResponse_location :: Lens' CreateDistributionWithTagsResponse (Maybe Text)

The fully qualified URI of the new distribution resource just created.

createDistributionWithTagsResponse_httpStatus :: Lens' CreateDistributionWithTagsResponse Int

The response's http status code.