Network.AWS.CloudFront.CreateInvalidation

Description

Create a new invalidation.

Creating a Request

data CreateInvalidation

The request to create an invalidation.

See: newCreateInvalidation smart constructor.

Constructors

CreateInvalidation'

Fields

distributionId :: Text

The distribution's id.

invalidationBatch :: InvalidationBatch

The batch information for the invalidation.

newCreateInvalidation

Arguments

:: Text
$sel:distributionId:CreateInvalidation'

-> InvalidationBatch
$sel:invalidationBatch:CreateInvalidation'

-> CreateInvalidation

Create a value of CreateInvalidation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:distributionId:CreateInvalidation', createInvalidation_distributionId - The distribution's id.

$sel:invalidationBatch:CreateInvalidation', createInvalidation_invalidationBatch - The batch information for the invalidation.

Request Lenses

createInvalidation_distributionId :: Lens' CreateInvalidation Text

The distribution's id.

createInvalidation_invalidationBatch :: Lens' CreateInvalidation InvalidationBatch

The batch information for the invalidation.

Destructuring the Response

data CreateInvalidationResponse

The returned result of the corresponding request.

See: newCreateInvalidationResponse smart constructor.

Constructors

CreateInvalidationResponse'

Fields

invalidation :: Maybe Invalidation

The invalidation's information.

location :: Maybe Text

The fully qualified URI of the distribution and invalidation batch request, including the Invalidation ID.

httpStatus :: Int

The response's http status code.

newCreateInvalidationResponse

Arguments

:: Int
$sel:httpStatus:CreateInvalidationResponse'

-> CreateInvalidationResponse

Create a value of CreateInvalidationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invalidation:CreateInvalidationResponse', createInvalidationResponse_invalidation - The invalidation's information.

$sel:location:CreateInvalidationResponse', createInvalidationResponse_location - The fully qualified URI of the distribution and invalidation batch request, including the Invalidation ID.

$sel:httpStatus:CreateInvalidationResponse', createInvalidationResponse_httpStatus - The response's http status code.

Response Lenses

createInvalidationResponse_invalidation :: Lens' CreateInvalidationResponse (Maybe Invalidation)

The invalidation's information.

createInvalidationResponse_location :: Lens' CreateInvalidationResponse (Maybe Text)

The fully qualified URI of the distribution and invalidation batch request, including the Invalidation ID.

createInvalidationResponse_httpStatus :: Lens' CreateInvalidationResponse Int

The response's http status code.