Network.AWS.CloudFront.CreateRealtimeLogConfig

Description

Creates a real-time log configuration.

After you create a real-time log configuration, you can attach it to one or more cache behaviors to send real-time log data to the specified Amazon Kinesis data stream.

For more information about real-time log configurations, see Real-time logs in the Amazon CloudFront Developer Guide.

Creating a Request

data CreateRealtimeLogConfig

See: newCreateRealtimeLogConfig smart constructor.

Constructors

CreateRealtimeLogConfig'

Fields

endPoints :: [EndPoint]

Contains information about the Amazon Kinesis data stream where you are sending real-time log data.

fields :: [Text]

A list of fields to include in each real-time log record.

For more information about fields, see Real-time log configuration fields in the Amazon CloudFront Developer Guide.

name :: Text

A unique name to identify this real-time log configuration.

samplingRate :: Integer

The sampling rate for this real-time log configuration. The sampling rate determines the percentage of viewer requests that are represented in the real-time log data. You must provide an integer between 1 and 100, inclusive.

newCreateRealtimeLogConfig

Arguments

:: Text
$sel:name:CreateRealtimeLogConfig'

-> Integer
$sel:samplingRate:CreateRealtimeLogConfig'

-> CreateRealtimeLogConfig

Create a value of CreateRealtimeLogConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endPoints:CreateRealtimeLogConfig', createRealtimeLogConfig_endPoints - Contains information about the Amazon Kinesis data stream where you are sending real-time log data.

$sel:fields:CreateRealtimeLogConfig', createRealtimeLogConfig_fields - A list of fields to include in each real-time log record.

For more information about fields, see Real-time log configuration fields in the Amazon CloudFront Developer Guide.

$sel:name:CreateRealtimeLogConfig', createRealtimeLogConfig_name - A unique name to identify this real-time log configuration.

$sel:samplingRate:CreateRealtimeLogConfig', createRealtimeLogConfig_samplingRate - The sampling rate for this real-time log configuration. The sampling rate determines the percentage of viewer requests that are represented in the real-time log data. You must provide an integer between 1 and 100, inclusive.

Request Lenses

createRealtimeLogConfig_endPoints :: Lens' CreateRealtimeLogConfig [EndPoint]

Contains information about the Amazon Kinesis data stream where you are sending real-time log data.

createRealtimeLogConfig_fields :: Lens' CreateRealtimeLogConfig [Text]

A list of fields to include in each real-time log record.

For more information about fields, see Real-time log configuration fields in the Amazon CloudFront Developer Guide.

createRealtimeLogConfig_name :: Lens' CreateRealtimeLogConfig Text

A unique name to identify this real-time log configuration.

createRealtimeLogConfig_samplingRate :: Lens' CreateRealtimeLogConfig Integer

The sampling rate for this real-time log configuration. The sampling rate determines the percentage of viewer requests that are represented in the real-time log data. You must provide an integer between 1 and 100, inclusive.

Destructuring the Response

data CreateRealtimeLogConfigResponse

See: newCreateRealtimeLogConfigResponse smart constructor.

Constructors

CreateRealtimeLogConfigResponse'

Fields

realtimeLogConfig :: Maybe RealtimeLogConfig

A real-time log configuration.

httpStatus :: Int

The response's http status code.

newCreateRealtimeLogConfigResponse

Arguments

:: Int
$sel:httpStatus:CreateRealtimeLogConfigResponse'

-> CreateRealtimeLogConfigResponse

Create a value of CreateRealtimeLogConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:realtimeLogConfig:CreateRealtimeLogConfigResponse', createRealtimeLogConfigResponse_realtimeLogConfig - A real-time log configuration.

$sel:httpStatus:CreateRealtimeLogConfigResponse', createRealtimeLogConfigResponse_httpStatus - The response's http status code.

Response Lenses

createRealtimeLogConfigResponse_realtimeLogConfig :: Lens' CreateRealtimeLogConfigResponse (Maybe RealtimeLogConfig)

A real-time log configuration.

createRealtimeLogConfigResponse_httpStatus :: Lens' CreateRealtimeLogConfigResponse Int

The response's http status code.