Network.AWS.CloudFront.CreateStreamingDistribution

Description

This API is deprecated. Amazon CloudFront is deprecating real-time messaging protocol (RTMP) distributions on December 31, 2020. For more information, read the announcement on the Amazon CloudFront discussion forum.

Creating a Request

data CreateStreamingDistribution

The request to create a new streaming distribution.

See: newCreateStreamingDistribution smart constructor.

Constructors

CreateStreamingDistribution'

Fields

streamingDistributionConfig :: StreamingDistributionConfig

The streaming distribution's configuration information.

newCreateStreamingDistribution

Arguments

:: StreamingDistributionConfig
$sel:streamingDistributionConfig:CreateStreamingDistribution'

-> CreateStreamingDistribution

Create a value of CreateStreamingDistribution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamingDistributionConfig:CreateStreamingDistribution', createStreamingDistribution_streamingDistributionConfig - The streaming distribution's configuration information.

Request Lenses

createStreamingDistribution_streamingDistributionConfig :: Lens' CreateStreamingDistribution StreamingDistributionConfig

The streaming distribution's configuration information.

Destructuring the Response

data CreateStreamingDistributionResponse

The returned result of the corresponding request.

See: newCreateStreamingDistributionResponse smart constructor.

Constructors

CreateStreamingDistributionResponse'

Fields

eTag :: Maybe Text

The current version of the streaming distribution created.

streamingDistribution :: Maybe StreamingDistribution

The streaming distribution's information.

location :: Maybe Text

The fully qualified URI of the new streaming distribution resource just created.

httpStatus :: Int

The response's http status code.

newCreateStreamingDistributionResponse

Arguments

:: Int
$sel:httpStatus:CreateStreamingDistributionResponse'

-> CreateStreamingDistributionResponse

Create a value of CreateStreamingDistributionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:CreateStreamingDistributionResponse', createStreamingDistributionResponse_eTag - The current version of the streaming distribution created.

$sel:streamingDistribution:CreateStreamingDistributionResponse', createStreamingDistributionResponse_streamingDistribution - The streaming distribution's information.

$sel:location:CreateStreamingDistributionResponse', createStreamingDistributionResponse_location - The fully qualified URI of the new streaming distribution resource just created.

$sel:httpStatus:CreateStreamingDistributionResponse', createStreamingDistributionResponse_httpStatus - The response's http status code.

Response Lenses

createStreamingDistributionResponse_eTag :: Lens' CreateStreamingDistributionResponse (Maybe Text)

The current version of the streaming distribution created.

createStreamingDistributionResponse_streamingDistribution :: Lens' CreateStreamingDistributionResponse (Maybe StreamingDistribution)

The streaming distribution's information.

createStreamingDistributionResponse_location :: Lens' CreateStreamingDistributionResponse (Maybe Text)

The fully qualified URI of the new streaming distribution resource just created.

createStreamingDistributionResponse_httpStatus :: Lens' CreateStreamingDistributionResponse Int

The response's http status code.