Network.AWS.CloudFront.CreateStreamingDistributionWithTags

Description

This API is deprecated. Amazon CloudFront is deprecating real-time messaging protocol (RTMP) distributions on December 31, 2020. For more information, read the announcement on the Amazon CloudFront discussion forum.

Creating a Request

data CreateStreamingDistributionWithTags

The request to create a new streaming distribution with tags.

See: newCreateStreamingDistributionWithTags smart constructor.

Constructors

CreateStreamingDistributionWithTags'

Fields

streamingDistributionConfigWithTags :: StreamingDistributionConfigWithTags

The streaming distribution's configuration information.

newCreateStreamingDistributionWithTags

Arguments

:: StreamingDistributionConfigWithTags
$sel:streamingDistributionConfigWithTags:CreateStreamingDistributionWithTags'

-> CreateStreamingDistributionWithTags

Create a value of CreateStreamingDistributionWithTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamingDistributionConfigWithTags:CreateStreamingDistributionWithTags', createStreamingDistributionWithTags_streamingDistributionConfigWithTags - The streaming distribution's configuration information.

Request Lenses

createStreamingDistributionWithTags_streamingDistributionConfigWithTags :: Lens' CreateStreamingDistributionWithTags StreamingDistributionConfigWithTags

The streaming distribution's configuration information.

Destructuring the Response

data CreateStreamingDistributionWithTagsResponse

The returned result of the corresponding request.

See: newCreateStreamingDistributionWithTagsResponse smart constructor.

Constructors

CreateStreamingDistributionWithTagsResponse'

Fields

eTag :: Maybe Text

The current version of the distribution created.

streamingDistribution :: Maybe StreamingDistribution

The streaming distribution's information.

location :: Maybe Text

The fully qualified URI of the new streaming distribution resource just created.

httpStatus :: Int

The response's http status code.

newCreateStreamingDistributionWithTagsResponse

Arguments

:: Int
$sel:httpStatus:CreateStreamingDistributionWithTagsResponse'

-> CreateStreamingDistributionWithTagsResponse

Create a value of CreateStreamingDistributionWithTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:CreateStreamingDistributionWithTagsResponse', createStreamingDistributionWithTagsResponse_eTag - The current version of the distribution created.

$sel:streamingDistribution:CreateStreamingDistributionWithTagsResponse', createStreamingDistributionWithTagsResponse_streamingDistribution - The streaming distribution's information.

$sel:location:CreateStreamingDistributionWithTagsResponse', createStreamingDistributionWithTagsResponse_location - The fully qualified URI of the new streaming distribution resource just created.

$sel:httpStatus:CreateStreamingDistributionWithTagsResponse', createStreamingDistributionWithTagsResponse_httpStatus - The response's http status code.

Response Lenses

createStreamingDistributionWithTagsResponse_eTag :: Lens' CreateStreamingDistributionWithTagsResponse (Maybe Text)

The current version of the distribution created.

createStreamingDistributionWithTagsResponse_streamingDistribution :: Lens' CreateStreamingDistributionWithTagsResponse (Maybe StreamingDistribution)

The streaming distribution's information.

createStreamingDistributionWithTagsResponse_location :: Lens' CreateStreamingDistributionWithTagsResponse (Maybe Text)

The fully qualified URI of the new streaming distribution resource just created.

createStreamingDistributionWithTagsResponse_httpStatus :: Lens' CreateStreamingDistributionWithTagsResponse Int

The response's http status code.