Network.AWS.CloudFront.DeleteDistribution

Description

Delete a distribution.

Creating a Request

data DeleteDistribution

This action deletes a web distribution. To delete a web distribution using the CloudFront API, perform the following steps.

To delete a web distribution using the CloudFront API:

Disable the web distribution

Submit a GET Distribution Config request to get the current configuration and the Etag header for the distribution.

Update the XML document that was returned in the response to your GET Distribution Config request to change the value of Enabled to false.

Submit a PUT Distribution Config request to update the configuration for your distribution. In the request body, include the XML document that you updated in Step 3. Set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GET Distribution Config request in Step 2.

Review the response to the PUT Distribution Config request to confirm that the distribution was successfully disabled.

Submit a GET Distribution request to confirm that your changes have propagated. When propagation is complete, the value of Status is Deployed.

Submit a DELETE Distribution request. Set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GET Distribution Config request in Step 6.

Review the response to your DELETE Distribution request to confirm that the distribution was successfully deleted.

For information about deleting a distribution using the CloudFront console, see Deleting a Distribution in the Amazon CloudFront Developer Guide.

See: newDeleteDistribution smart constructor.

Constructors

DeleteDistribution'

Fields

ifMatch :: Maybe Text

The value of the ETag header that you received when you disabled the distribution. For example: E2QWRUHAPOMQZL.

id :: Text

The distribution ID.

newDeleteDistribution

Arguments

:: Text
$sel:id:DeleteDistribution'

-> DeleteDistribution

Create a value of DeleteDistribution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifMatch:DeleteDistribution', deleteDistribution_ifMatch - The value of the ETag header that you received when you disabled the distribution. For example: E2QWRUHAPOMQZL.

$sel:id:DeleteDistribution', deleteDistribution_id - The distribution ID.

Request Lenses

deleteDistribution_ifMatch :: Lens' DeleteDistribution (Maybe Text)

The value of the ETag header that you received when you disabled the distribution. For example: E2QWRUHAPOMQZL.

deleteDistribution_id :: Lens' DeleteDistribution Text

The distribution ID.

Destructuring the Response

data DeleteDistributionResponse

See: newDeleteDistributionResponse smart constructor.

Constructors

DeleteDistributionResponse'

newDeleteDistributionResponse :: DeleteDistributionResponse

Create a value of DeleteDistributionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.