Network.AWS.CloudFront.DeleteFieldLevelEncryptionConfig

Description

Remove a field-level encryption configuration.

Creating a Request

data DeleteFieldLevelEncryptionConfig

See: newDeleteFieldLevelEncryptionConfig smart constructor.

Constructors

DeleteFieldLevelEncryptionConfig'

Fields

ifMatch :: Maybe Text

The value of the ETag header that you received when retrieving the configuration identity to delete. For example: E2QWRUHAPOMQZL.

id :: Text

The ID of the configuration you want to delete from CloudFront.

newDeleteFieldLevelEncryptionConfig

Arguments

:: Text
$sel:id:DeleteFieldLevelEncryptionConfig'

-> DeleteFieldLevelEncryptionConfig

Create a value of DeleteFieldLevelEncryptionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifMatch:DeleteFieldLevelEncryptionConfig', deleteFieldLevelEncryptionConfig_ifMatch - The value of the ETag header that you received when retrieving the configuration identity to delete. For example: E2QWRUHAPOMQZL.

$sel:id:DeleteFieldLevelEncryptionConfig', deleteFieldLevelEncryptionConfig_id - The ID of the configuration you want to delete from CloudFront.

Request Lenses

deleteFieldLevelEncryptionConfig_ifMatch :: Lens' DeleteFieldLevelEncryptionConfig (Maybe Text)

The value of the ETag header that you received when retrieving the configuration identity to delete. For example: E2QWRUHAPOMQZL.

deleteFieldLevelEncryptionConfig_id :: Lens' DeleteFieldLevelEncryptionConfig Text

The ID of the configuration you want to delete from CloudFront.

Destructuring the Response

data DeleteFieldLevelEncryptionConfigResponse

See: newDeleteFieldLevelEncryptionConfigResponse smart constructor.

Constructors

DeleteFieldLevelEncryptionConfigResponse'

newDeleteFieldLevelEncryptionConfigResponse :: DeleteFieldLevelEncryptionConfigResponse

Create a value of DeleteFieldLevelEncryptionConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.