Network.AWS.CloudFront.DeleteStreamingDistribution

Description

Delete a streaming distribution. To delete an RTMP distribution using the CloudFront API, perform the following steps.

To delete an RTMP distribution using the CloudFront API:

Disable the RTMP distribution.

Submit a GET Streaming Distribution Config request to get the current configuration and the Etag header for the distribution.

Update the XML document that was returned in the response to your GET Streaming Distribution Config request to change the value of Enabled to false.

Submit a PUT Streaming Distribution Config request to update the configuration for your distribution. In the request body, include the XML document that you updated in Step 3. Then set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GET Streaming Distribution Config request in Step 2.

Review the response to the PUT Streaming Distribution Config request to confirm that the distribution was successfully disabled.

Submit a GET Streaming Distribution Config request to confirm that your changes have propagated. When propagation is complete, the value of Status is Deployed.

Submit a DELETE Streaming Distribution request. Set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GET Streaming Distribution Config request in Step 2.

Review the response to your DELETE Streaming Distribution request to confirm that the distribution was successfully deleted.

For information about deleting a distribution using the CloudFront console, see Deleting a Distribution in the Amazon CloudFront Developer Guide.

Creating a Request

data DeleteStreamingDistribution

The request to delete a streaming distribution.

See: newDeleteStreamingDistribution smart constructor.

Constructors

DeleteStreamingDistribution'

Fields

ifMatch :: Maybe Text

The value of the ETag header that you received when you disabled the streaming distribution. For example: E2QWRUHAPOMQZL.

id :: Text

The distribution ID.

newDeleteStreamingDistribution

Arguments

:: Text
$sel:id:DeleteStreamingDistribution'

-> DeleteStreamingDistribution

Create a value of DeleteStreamingDistribution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifMatch:DeleteStreamingDistribution', deleteStreamingDistribution_ifMatch - The value of the ETag header that you received when you disabled the streaming distribution. For example: E2QWRUHAPOMQZL.

$sel:id:DeleteStreamingDistribution', deleteStreamingDistribution_id - The distribution ID.

Request Lenses

deleteStreamingDistribution_ifMatch :: Lens' DeleteStreamingDistribution (Maybe Text)

The value of the ETag header that you received when you disabled the streaming distribution. For example: E2QWRUHAPOMQZL.

deleteStreamingDistribution_id :: Lens' DeleteStreamingDistribution Text

The distribution ID.

Destructuring the Response

data DeleteStreamingDistributionResponse

See: newDeleteStreamingDistributionResponse smart constructor.

Constructors

DeleteStreamingDistributionResponse'

newDeleteStreamingDistributionResponse :: DeleteStreamingDistributionResponse

Create a value of DeleteStreamingDistributionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.