Network.AWS.CloudFront.GetCachePolicy

Description

Gets a cache policy, including the following metadata:

The policy’s identifier.

The date and time when the policy was last modified.

To get a cache policy, you must provide the policy’s identifier. If the cache policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the cache policy is not attached to a cache behavior, you can get the identifier using ListCachePolicies.

Creating a Request

data GetCachePolicy

See: newGetCachePolicy smart constructor.

Constructors

GetCachePolicy'

Fields

id :: Text

The unique identifier for the cache policy. If the cache policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the cache policy is not attached to a cache behavior, you can get the identifier using ListCachePolicies.

newGetCachePolicy

Arguments

:: Text
$sel:id:GetCachePolicy'

-> GetCachePolicy

Create a value of GetCachePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetCachePolicy', getCachePolicy_id - The unique identifier for the cache policy. If the cache policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the cache policy is not attached to a cache behavior, you can get the identifier using ListCachePolicies.

Request Lenses

getCachePolicy_id :: Lens' GetCachePolicy Text

The unique identifier for the cache policy. If the cache policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the cache policy is not attached to a cache behavior, you can get the identifier using ListCachePolicies.

Destructuring the Response

data GetCachePolicyResponse

See: newGetCachePolicyResponse smart constructor.

Constructors

GetCachePolicyResponse'

Fields

eTag :: Maybe Text

The current version of the cache policy.

cachePolicy :: Maybe CachePolicy

The cache policy.

httpStatus :: Int

The response's http status code.

newGetCachePolicyResponse

Arguments

:: Int
$sel:httpStatus:GetCachePolicyResponse'

-> GetCachePolicyResponse

Create a value of GetCachePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:GetCachePolicyResponse', getCachePolicyResponse_eTag - The current version of the cache policy.

$sel:cachePolicy:GetCachePolicyResponse', getCachePolicyResponse_cachePolicy - The cache policy.

$sel:httpStatus:GetCachePolicyResponse', getCachePolicyResponse_httpStatus - The response's http status code.

Response Lenses

getCachePolicyResponse_eTag :: Lens' GetCachePolicyResponse (Maybe Text)

The current version of the cache policy.

getCachePolicyResponse_cachePolicy :: Lens' GetCachePolicyResponse (Maybe CachePolicy)

The cache policy.

getCachePolicyResponse_httpStatus :: Lens' GetCachePolicyResponse Int

The response's http status code.