Network.AWS.CloudFront.GetFieldLevelEncryptionConfig

Description

Get the field-level encryption configuration information.

Creating a Request

data GetFieldLevelEncryptionConfig

See: newGetFieldLevelEncryptionConfig smart constructor.

Constructors

GetFieldLevelEncryptionConfig'

Fields

id :: Text

Request the ID for the field-level encryption configuration information.

newGetFieldLevelEncryptionConfig

Arguments

:: Text
$sel:id:GetFieldLevelEncryptionConfig'

-> GetFieldLevelEncryptionConfig

Create a value of GetFieldLevelEncryptionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetFieldLevelEncryptionConfig', getFieldLevelEncryptionConfig_id - Request the ID for the field-level encryption configuration information.

Request Lenses

getFieldLevelEncryptionConfig_id :: Lens' GetFieldLevelEncryptionConfig Text

Request the ID for the field-level encryption configuration information.

Destructuring the Response

data GetFieldLevelEncryptionConfigResponse

See: newGetFieldLevelEncryptionConfigResponse smart constructor.

Constructors

GetFieldLevelEncryptionConfigResponse'

Fields

eTag :: Maybe Text

The current version of the field level encryption configuration. For example: E2QWRUHAPOMQZL.

fieldLevelEncryptionConfig :: Maybe FieldLevelEncryptionConfig

Return the field-level encryption configuration information.

httpStatus :: Int

The response's http status code.

newGetFieldLevelEncryptionConfigResponse

Arguments

:: Int
$sel:httpStatus:GetFieldLevelEncryptionConfigResponse'

-> GetFieldLevelEncryptionConfigResponse

Create a value of GetFieldLevelEncryptionConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:GetFieldLevelEncryptionConfigResponse', getFieldLevelEncryptionConfigResponse_eTag - The current version of the field level encryption configuration. For example: E2QWRUHAPOMQZL.

$sel:fieldLevelEncryptionConfig:GetFieldLevelEncryptionConfigResponse', getFieldLevelEncryptionConfigResponse_fieldLevelEncryptionConfig - Return the field-level encryption configuration information.

$sel:httpStatus:GetFieldLevelEncryptionConfigResponse', getFieldLevelEncryptionConfigResponse_httpStatus - The response's http status code.

Response Lenses

getFieldLevelEncryptionConfigResponse_eTag :: Lens' GetFieldLevelEncryptionConfigResponse (Maybe Text)

The current version of the field level encryption configuration. For example: E2QWRUHAPOMQZL.

getFieldLevelEncryptionConfigResponse_fieldLevelEncryptionConfig :: Lens' GetFieldLevelEncryptionConfigResponse (Maybe FieldLevelEncryptionConfig)

Return the field-level encryption configuration information.

getFieldLevelEncryptionConfigResponse_httpStatus :: Lens' GetFieldLevelEncryptionConfigResponse Int

The response's http status code.