Network.AWS.CloudFront.GetFunction

Description

Gets the code of a CloudFront function. To get configuration information and metadata about a function, use DescribeFunction.

To get a function’s code, you must provide the function’s name and stage. To get these values, you can use ListFunctions.

Creating a Request

data GetFunction

See: newGetFunction smart constructor.

Constructors

GetFunction'

Fields

stage :: Maybe FunctionStage

The function’s stage, either DEVELOPMENT or LIVE.

name :: Text

The name of the function whose code you are getting.

newGetFunction

Arguments

:: Text
$sel:name:GetFunction'

-> GetFunction

Create a value of GetFunction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stage:GetFunction', getFunction_stage - The function’s stage, either DEVELOPMENT or LIVE.

$sel:name:GetFunction', getFunction_name - The name of the function whose code you are getting.

Request Lenses

getFunction_stage :: Lens' GetFunction (Maybe FunctionStage)

The function’s stage, either DEVELOPMENT or LIVE.

getFunction_name :: Lens' GetFunction Text

The name of the function whose code you are getting.

Destructuring the Response

data GetFunctionResponse

See: newGetFunctionResponse smart constructor.

Constructors

GetFunctionResponse'

Fields

eTag :: Maybe Text

The version identifier for the current version of the CloudFront function.

contentType :: Maybe Text

The content type (media type) of the response.

functionCode :: Maybe (Sensitive ByteString)

The function code of a CloudFront function.

httpStatus :: Int

The response's http status code.

newGetFunctionResponse

Arguments

:: Int
$sel:httpStatus:GetFunctionResponse'

-> GetFunctionResponse

Create a value of GetFunctionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:GetFunctionResponse', getFunctionResponse_eTag - The version identifier for the current version of the CloudFront function.

$sel:contentType:GetFunctionResponse', getFunctionResponse_contentType - The content type (media type) of the response.

$sel:functionCode:GetFunctionResponse', getFunctionResponse_functionCode - The function code of a CloudFront function.

$sel:httpStatus:GetFunctionResponse', getFunctionResponse_httpStatus - The response's http status code.

Response Lenses

getFunctionResponse_eTag :: Lens' GetFunctionResponse (Maybe Text)

The version identifier for the current version of the CloudFront function.

getFunctionResponse_contentType :: Lens' GetFunctionResponse (Maybe Text)

The content type (media type) of the response.

getFunctionResponse_functionCode :: Lens' GetFunctionResponse (Maybe ByteString)

The function code of a CloudFront function.

getFunctionResponse_httpStatus :: Lens' GetFunctionResponse Int

The response's http status code.