Network.AWS.CloudFront.GetStreamingDistribution

Description

Gets information about a specified RTMP distribution, including the distribution configuration.

Creating a Request

data GetStreamingDistribution

The request to get a streaming distribution's information.

See: newGetStreamingDistribution smart constructor.

Constructors

GetStreamingDistribution'

Fields

id :: Text

The streaming distribution's ID.

newGetStreamingDistribution

Arguments

:: Text
$sel:id:GetStreamingDistribution'

-> GetStreamingDistribution

Create a value of GetStreamingDistribution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetStreamingDistribution', getStreamingDistribution_id - The streaming distribution's ID.

Request Lenses

getStreamingDistribution_id :: Lens' GetStreamingDistribution Text

The streaming distribution's ID.

Destructuring the Response

data GetStreamingDistributionResponse

The returned result of the corresponding request.

See: newGetStreamingDistributionResponse smart constructor.

Constructors

GetStreamingDistributionResponse'

Fields

eTag :: Maybe Text

The current version of the streaming distribution's information. For example: E2QWRUHAPOMQZL.

streamingDistribution :: Maybe StreamingDistribution

The streaming distribution's information.

httpStatus :: Int

The response's http status code.

newGetStreamingDistributionResponse

Arguments

:: Int
$sel:httpStatus:GetStreamingDistributionResponse'

-> GetStreamingDistributionResponse

Create a value of GetStreamingDistributionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:GetStreamingDistributionResponse', getStreamingDistributionResponse_eTag - The current version of the streaming distribution's information. For example: E2QWRUHAPOMQZL.

$sel:streamingDistribution:GetStreamingDistributionResponse', getStreamingDistributionResponse_streamingDistribution - The streaming distribution's information.

$sel:httpStatus:GetStreamingDistributionResponse', getStreamingDistributionResponse_httpStatus - The response's http status code.

Response Lenses

getStreamingDistributionResponse_eTag :: Lens' GetStreamingDistributionResponse (Maybe Text)

The current version of the streaming distribution's information. For example: E2QWRUHAPOMQZL.

getStreamingDistributionResponse_streamingDistribution :: Lens' GetStreamingDistributionResponse (Maybe StreamingDistribution)

The streaming distribution's information.

getStreamingDistributionResponse_httpStatus :: Lens' GetStreamingDistributionResponse Int

The response's http status code.