Network.AWS.CloudFront.GetStreamingDistributionConfig

Description

Get the configuration information about a streaming distribution.

Creating a Request

data GetStreamingDistributionConfig

To request to get a streaming distribution configuration.

See: newGetStreamingDistributionConfig smart constructor.

Constructors

GetStreamingDistributionConfig'

Fields

id :: Text

The streaming distribution's ID.

newGetStreamingDistributionConfig

Arguments

:: Text
$sel:id:GetStreamingDistributionConfig'

-> GetStreamingDistributionConfig

Create a value of GetStreamingDistributionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetStreamingDistributionConfig', getStreamingDistributionConfig_id - The streaming distribution's ID.

Request Lenses

getStreamingDistributionConfig_id :: Lens' GetStreamingDistributionConfig Text

The streaming distribution's ID.

Destructuring the Response

data GetStreamingDistributionConfigResponse

The returned result of the corresponding request.

See: newGetStreamingDistributionConfigResponse smart constructor.

Constructors

GetStreamingDistributionConfigResponse'

Fields

streamingDistributionConfig :: Maybe StreamingDistributionConfig

The streaming distribution's configuration information.

eTag :: Maybe Text

The current version of the configuration. For example: E2QWRUHAPOMQZL.

httpStatus :: Int

The response's http status code.

newGetStreamingDistributionConfigResponse

Arguments

:: Int
$sel:httpStatus:GetStreamingDistributionConfigResponse'

-> GetStreamingDistributionConfigResponse

Create a value of GetStreamingDistributionConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamingDistributionConfig:GetStreamingDistributionConfigResponse', getStreamingDistributionConfigResponse_streamingDistributionConfig - The streaming distribution's configuration information.

$sel:eTag:GetStreamingDistributionConfigResponse', getStreamingDistributionConfigResponse_eTag - The current version of the configuration. For example: E2QWRUHAPOMQZL.

$sel:httpStatus:GetStreamingDistributionConfigResponse', getStreamingDistributionConfigResponse_httpStatus - The response's http status code.

Response Lenses

getStreamingDistributionConfigResponse_streamingDistributionConfig :: Lens' GetStreamingDistributionConfigResponse (Maybe StreamingDistributionConfig)

The streaming distribution's configuration information.

getStreamingDistributionConfigResponse_eTag :: Lens' GetStreamingDistributionConfigResponse (Maybe Text)

The current version of the configuration. For example: E2QWRUHAPOMQZL.

getStreamingDistributionConfigResponse_httpStatus :: Lens' GetStreamingDistributionConfigResponse Int

The response's http status code.