Network.AWS.CloudFront.ListConflictingAliases

Description

Gets a list of aliases (also called CNAMEs or alternate domain names) that conflict or overlap with the provided alias, and the associated CloudFront distributions and Amazon Web Services accounts for each conflicting alias. In the returned list, the distribution and account IDs are partially hidden, which allows you to identify the distributions and accounts that you own, but helps to protect the information of ones that you don’t own.

Use this operation to find aliases that are in use in CloudFront that conflict or overlap with the provided alias. For example, if you provide www.example.com as input, the returned list can include www.example.com and the overlapping wildcard alternate domain name (*.example.com), if they exist. If you provide *.example.com as input, the returned list can include *.example.com and any alternate domain names covered by that wildcard (for example, www.example.com, test.example.com, dev.example.com, and so on), if they exist.

To list conflicting aliases, you provide the alias to search and the ID of a distribution in your account that has an attached SSL/TLS certificate that includes the provided alias. For more information, including how to set up the distribution and certificate, see Moving an alternate domain name to a different distribution in the Amazon CloudFront Developer Guide.

You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

Creating a Request

data ListConflictingAliases

See: newListConflictingAliases smart constructor.

Constructors

ListConflictingAliases'

Fields

maxItems :: Maybe Int

The maximum number of conflicting aliases that you want in the response.

marker :: Maybe Text

Use this field when paginating results to indicate where to begin in the list of conflicting aliases. The response includes conflicting aliases in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

distributionId :: Text

The ID of a distribution in your account that has an attached SSL/TLS certificate that includes the provided alias.

alias :: Text

The alias (also called a CNAME) to search for conflicting aliases.

newListConflictingAliases

Arguments

:: Text
$sel:distributionId:ListConflictingAliases'

-> Text
$sel:alias:ListConflictingAliases'

-> ListConflictingAliases

Create a value of ListConflictingAliases with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxItems:ListConflictingAliases', listConflictingAliases_maxItems - The maximum number of conflicting aliases that you want in the response.

$sel:marker:ListConflictingAliases', listConflictingAliases_marker - Use this field when paginating results to indicate where to begin in the list of conflicting aliases. The response includes conflicting aliases in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

$sel:distributionId:ListConflictingAliases', listConflictingAliases_distributionId - The ID of a distribution in your account that has an attached SSL/TLS certificate that includes the provided alias.

$sel:alias:ListConflictingAliases', listConflictingAliases_alias - The alias (also called a CNAME) to search for conflicting aliases.

Request Lenses

listConflictingAliases_maxItems :: Lens' ListConflictingAliases (Maybe Int)

The maximum number of conflicting aliases that you want in the response.

listConflictingAliases_marker :: Lens' ListConflictingAliases (Maybe Text)

Use this field when paginating results to indicate where to begin in the list of conflicting aliases. The response includes conflicting aliases in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

listConflictingAliases_distributionId :: Lens' ListConflictingAliases Text

The ID of a distribution in your account that has an attached SSL/TLS certificate that includes the provided alias.

listConflictingAliases_alias :: Lens' ListConflictingAliases Text

The alias (also called a CNAME) to search for conflicting aliases.

Destructuring the Response

data ListConflictingAliasesResponse

See: newListConflictingAliasesResponse smart constructor.

Constructors

ListConflictingAliasesResponse'

Fields

conflictingAliasesList :: Maybe ConflictingAliasesList

A list of conflicting aliases.

httpStatus :: Int

The response's http status code.

newListConflictingAliasesResponse

Arguments

:: Int
$sel:httpStatus:ListConflictingAliasesResponse'

-> ListConflictingAliasesResponse

Create a value of ListConflictingAliasesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conflictingAliasesList:ListConflictingAliasesResponse', listConflictingAliasesResponse_conflictingAliasesList - A list of conflicting aliases.

$sel:httpStatus:ListConflictingAliasesResponse', listConflictingAliasesResponse_httpStatus - The response's http status code.

Response Lenses

listConflictingAliasesResponse_conflictingAliasesList :: Lens' ListConflictingAliasesResponse (Maybe ConflictingAliasesList)

A list of conflicting aliases.

listConflictingAliasesResponse_httpStatus :: Lens' ListConflictingAliasesResponse Int

The response's http status code.