Network.AWS.CloudFront.ListDistributionsByRealtimeLogConfig

Description

Gets a list of distributions that have a cache behavior that’s associated with the specified real-time log configuration.

You can specify the real-time log configuration by its name or its Amazon Resource Name (ARN). You must provide at least one. If you provide both, CloudFront uses the name to identify the real-time log configuration to list distributions for.

You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

Creating a Request

data ListDistributionsByRealtimeLogConfig

See: newListDistributionsByRealtimeLogConfig smart constructor.

Constructors

ListDistributionsByRealtimeLogConfig'

Fields

realtimeLogConfigName :: Maybe Text

The name of the real-time log configuration whose associated distributions you want to list.

realtimeLogConfigArn :: Maybe Text

The Amazon Resource Name (ARN) of the real-time log configuration whose associated distributions you want to list.

marker :: Maybe Text

Use this field when paginating results to indicate where to begin in your list of distributions. The response includes distributions in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

maxItems :: Maybe Text

The maximum number of distributions that you want in the response.

newListDistributionsByRealtimeLogConfig :: ListDistributionsByRealtimeLogConfig

Create a value of ListDistributionsByRealtimeLogConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:realtimeLogConfigName:ListDistributionsByRealtimeLogConfig', listDistributionsByRealtimeLogConfig_realtimeLogConfigName - The name of the real-time log configuration whose associated distributions you want to list.

$sel:realtimeLogConfigArn:ListDistributionsByRealtimeLogConfig', listDistributionsByRealtimeLogConfig_realtimeLogConfigArn - The Amazon Resource Name (ARN) of the real-time log configuration whose associated distributions you want to list.

$sel:marker:ListDistributionsByRealtimeLogConfig', listDistributionsByRealtimeLogConfig_marker - Use this field when paginating results to indicate where to begin in your list of distributions. The response includes distributions in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

$sel:maxItems:ListDistributionsByRealtimeLogConfig', listDistributionsByRealtimeLogConfig_maxItems - The maximum number of distributions that you want in the response.

Request Lenses

listDistributionsByRealtimeLogConfig_realtimeLogConfigName :: Lens' ListDistributionsByRealtimeLogConfig (Maybe Text)

The name of the real-time log configuration whose associated distributions you want to list.

listDistributionsByRealtimeLogConfig_realtimeLogConfigArn :: Lens' ListDistributionsByRealtimeLogConfig (Maybe Text)

The Amazon Resource Name (ARN) of the real-time log configuration whose associated distributions you want to list.

listDistributionsByRealtimeLogConfig_marker :: Lens' ListDistributionsByRealtimeLogConfig (Maybe Text)

Use this field when paginating results to indicate where to begin in your list of distributions. The response includes distributions in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

listDistributionsByRealtimeLogConfig_maxItems :: Lens' ListDistributionsByRealtimeLogConfig (Maybe Text)

The maximum number of distributions that you want in the response.

Destructuring the Response

data ListDistributionsByRealtimeLogConfigResponse

See: newListDistributionsByRealtimeLogConfigResponse smart constructor.

Constructors

ListDistributionsByRealtimeLogConfigResponse'

Fields

distributionList :: Maybe DistributionList

httpStatus :: Int

The response's http status code.

newListDistributionsByRealtimeLogConfigResponse

Arguments

:: Int
$sel:httpStatus:ListDistributionsByRealtimeLogConfigResponse'

-> ListDistributionsByRealtimeLogConfigResponse

Create a value of ListDistributionsByRealtimeLogConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:distributionList:ListDistributionsByRealtimeLogConfigResponse', listDistributionsByRealtimeLogConfigResponse_distributionList - Undocumented member.

$sel:httpStatus:ListDistributionsByRealtimeLogConfigResponse', listDistributionsByRealtimeLogConfigResponse_httpStatus - The response's http status code.

Response Lenses

listDistributionsByRealtimeLogConfigResponse_distributionList :: Lens' ListDistributionsByRealtimeLogConfigResponse (Maybe DistributionList)

Undocumented member.

listDistributionsByRealtimeLogConfigResponse_httpStatus :: Lens' ListDistributionsByRealtimeLogConfigResponse Int

The response's http status code.