Network.AWS.CloudFront.ListDistributionsByWebACLId

Description

List the distributions that are associated with a specified WAF web ACL.

Creating a Request

data ListDistributionsByWebACLId

The request to list distributions that are associated with a specified WAF web ACL.

See: newListDistributionsByWebACLId smart constructor.

Constructors

ListDistributionsByWebACLId'

Fields

marker :: Maybe Text

Use Marker and MaxItems to control pagination of results. If you have more than MaxItems distributions that satisfy the request, the response includes a NextMarker element. To get the next page of results, submit another request. For the value of Marker, specify the value of NextMarker from the last response. (For the first request, omit Marker.)

maxItems :: Maybe Text

The maximum number of distributions that you want CloudFront to return in the response body. The maximum and default values are both 100.

webACLId :: Text

The ID of the WAF web ACL that you want to list the associated distributions. If you specify "null" for the ID, the request returns a list of the distributions that aren't associated with a web ACL.

newListDistributionsByWebACLId

Arguments

:: Text
$sel:webACLId:ListDistributionsByWebACLId'

-> ListDistributionsByWebACLId

Create a value of ListDistributionsByWebACLId with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListDistributionsByWebACLId', listDistributionsByWebACLId_marker - Use Marker and MaxItems to control pagination of results. If you have more than MaxItems distributions that satisfy the request, the response includes a NextMarker element. To get the next page of results, submit another request. For the value of Marker, specify the value of NextMarker from the last response. (For the first request, omit Marker.)

$sel:maxItems:ListDistributionsByWebACLId', listDistributionsByWebACLId_maxItems - The maximum number of distributions that you want CloudFront to return in the response body. The maximum and default values are both 100.

$sel:webACLId:ListDistributionsByWebACLId', listDistributionsByWebACLId_webACLId - The ID of the WAF web ACL that you want to list the associated distributions. If you specify "null" for the ID, the request returns a list of the distributions that aren't associated with a web ACL.

Request Lenses

listDistributionsByWebACLId_marker :: Lens' ListDistributionsByWebACLId (Maybe Text)

Use Marker and MaxItems to control pagination of results. If you have more than MaxItems distributions that satisfy the request, the response includes a NextMarker element. To get the next page of results, submit another request. For the value of Marker, specify the value of NextMarker from the last response. (For the first request, omit Marker.)

listDistributionsByWebACLId_maxItems :: Lens' ListDistributionsByWebACLId (Maybe Text)

The maximum number of distributions that you want CloudFront to return in the response body. The maximum and default values are both 100.

listDistributionsByWebACLId_webACLId :: Lens' ListDistributionsByWebACLId Text

The ID of the WAF web ACL that you want to list the associated distributions. If you specify "null" for the ID, the request returns a list of the distributions that aren't associated with a web ACL.

Destructuring the Response

data ListDistributionsByWebACLIdResponse

The response to a request to list the distributions that are associated with a specified WAF web ACL.

See: newListDistributionsByWebACLIdResponse smart constructor.

Constructors

ListDistributionsByWebACLIdResponse'

Fields

distributionList :: Maybe DistributionList

The DistributionList type.

httpStatus :: Int

The response's http status code.

newListDistributionsByWebACLIdResponse

Arguments

:: Int
$sel:httpStatus:ListDistributionsByWebACLIdResponse'

-> ListDistributionsByWebACLIdResponse

Create a value of ListDistributionsByWebACLIdResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:distributionList:ListDistributionsByWebACLIdResponse', listDistributionsByWebACLIdResponse_distributionList - The DistributionList type.

$sel:httpStatus:ListDistributionsByWebACLIdResponse', listDistributionsByWebACLIdResponse_httpStatus - The response's http status code.

Response Lenses

listDistributionsByWebACLIdResponse_distributionList :: Lens' ListDistributionsByWebACLIdResponse (Maybe DistributionList)

The DistributionList type.

listDistributionsByWebACLIdResponse_httpStatus :: Lens' ListDistributionsByWebACLIdResponse Int

The response's http status code.