Network.AWS.CloudFront.ListFieldLevelEncryptionProfiles

Description

Request a list of field-level encryption profiles that have been created in CloudFront for this account.

Creating a Request

data ListFieldLevelEncryptionProfiles

See: newListFieldLevelEncryptionProfiles smart constructor.

Constructors

ListFieldLevelEncryptionProfiles'

Fields

maxItems :: Maybe Text

The maximum number of field-level encryption profiles you want in the response body.

marker :: Maybe Text

Use this when paginating results to indicate where to begin in your list of profiles. The results include profiles in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last profile on that page).

newListFieldLevelEncryptionProfiles :: ListFieldLevelEncryptionProfiles

Create a value of ListFieldLevelEncryptionProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxItems:ListFieldLevelEncryptionProfiles', listFieldLevelEncryptionProfiles_maxItems - The maximum number of field-level encryption profiles you want in the response body.

$sel:marker:ListFieldLevelEncryptionProfiles', listFieldLevelEncryptionProfiles_marker - Use this when paginating results to indicate where to begin in your list of profiles. The results include profiles in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last profile on that page).

Request Lenses

listFieldLevelEncryptionProfiles_maxItems :: Lens' ListFieldLevelEncryptionProfiles (Maybe Text)

The maximum number of field-level encryption profiles you want in the response body.

listFieldLevelEncryptionProfiles_marker :: Lens' ListFieldLevelEncryptionProfiles (Maybe Text)

Use this when paginating results to indicate where to begin in your list of profiles. The results include profiles in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last profile on that page).

Destructuring the Response

data ListFieldLevelEncryptionProfilesResponse

See: newListFieldLevelEncryptionProfilesResponse smart constructor.

Constructors

ListFieldLevelEncryptionProfilesResponse'

Fields

fieldLevelEncryptionProfileList :: Maybe FieldLevelEncryptionProfileList

Returns a list of the field-level encryption profiles that have been created in CloudFront for this account.

httpStatus :: Int

The response's http status code.

newListFieldLevelEncryptionProfilesResponse

Arguments

:: Int
$sel:httpStatus:ListFieldLevelEncryptionProfilesResponse'

-> ListFieldLevelEncryptionProfilesResponse

Create a value of ListFieldLevelEncryptionProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fieldLevelEncryptionProfileList:ListFieldLevelEncryptionProfilesResponse', listFieldLevelEncryptionProfilesResponse_fieldLevelEncryptionProfileList - Returns a list of the field-level encryption profiles that have been created in CloudFront for this account.

$sel:httpStatus:ListFieldLevelEncryptionProfilesResponse', listFieldLevelEncryptionProfilesResponse_httpStatus - The response's http status code.

Response Lenses

listFieldLevelEncryptionProfilesResponse_fieldLevelEncryptionProfileList :: Lens' ListFieldLevelEncryptionProfilesResponse (Maybe FieldLevelEncryptionProfileList)

Returns a list of the field-level encryption profiles that have been created in CloudFront for this account.

listFieldLevelEncryptionProfilesResponse_httpStatus :: Lens' ListFieldLevelEncryptionProfilesResponse Int

The response's http status code.