Network.AWS.CloudFront.ListRealtimeLogConfigs

Description

Gets a list of real-time log configurations.

You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

Creating a Request

data ListRealtimeLogConfigs

See: newListRealtimeLogConfigs smart constructor.

Constructors

ListRealtimeLogConfigs'

Fields

maxItems :: Maybe Text

The maximum number of real-time log configurations that you want in the response.

marker :: Maybe Text

Use this field when paginating results to indicate where to begin in your list of real-time log configurations. The response includes real-time log configurations in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

newListRealtimeLogConfigs :: ListRealtimeLogConfigs

Create a value of ListRealtimeLogConfigs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxItems:ListRealtimeLogConfigs', listRealtimeLogConfigs_maxItems - The maximum number of real-time log configurations that you want in the response.

$sel:marker:ListRealtimeLogConfigs', listRealtimeLogConfigs_marker - Use this field when paginating results to indicate where to begin in your list of real-time log configurations. The response includes real-time log configurations in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

Request Lenses

listRealtimeLogConfigs_maxItems :: Lens' ListRealtimeLogConfigs (Maybe Text)

The maximum number of real-time log configurations that you want in the response.

listRealtimeLogConfigs_marker :: Lens' ListRealtimeLogConfigs (Maybe Text)

Use this field when paginating results to indicate where to begin in your list of real-time log configurations. The response includes real-time log configurations in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

Destructuring the Response

data ListRealtimeLogConfigsResponse

See: newListRealtimeLogConfigsResponse smart constructor.

Constructors

ListRealtimeLogConfigsResponse'

Fields

realtimeLogConfigs :: Maybe RealtimeLogConfigs

A list of real-time log configurations.

httpStatus :: Int

The response's http status code.

newListRealtimeLogConfigsResponse

Arguments

:: Int
$sel:httpStatus:ListRealtimeLogConfigsResponse'

-> ListRealtimeLogConfigsResponse

Create a value of ListRealtimeLogConfigsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:realtimeLogConfigs:ListRealtimeLogConfigsResponse', listRealtimeLogConfigsResponse_realtimeLogConfigs - A list of real-time log configurations.

$sel:httpStatus:ListRealtimeLogConfigsResponse', listRealtimeLogConfigsResponse_httpStatus - The response's http status code.

Response Lenses

listRealtimeLogConfigsResponse_realtimeLogConfigs :: Lens' ListRealtimeLogConfigsResponse (Maybe RealtimeLogConfigs)

A list of real-time log configurations.

listRealtimeLogConfigsResponse_httpStatus :: Lens' ListRealtimeLogConfigsResponse Int

The response's http status code.