Network.AWS.CloudFront.PublishFunction

Description

Publishes a CloudFront function by copying the function code from the DEVELOPMENT stage to LIVE. This automatically updates all cache behaviors that are using this function to use the newly published copy in the LIVE stage.

When a function is published to the LIVE stage, you can attach the function to a distribution’s cache behavior, using the function’s Amazon Resource Name (ARN).

To publish a function, you must provide the function’s name and version (ETag value). To get these values, you can use ListFunctions and DescribeFunction.

Creating a Request

data PublishFunction

See: newPublishFunction smart constructor.

Constructors

PublishFunction'

Fields

name :: Text

The name of the function that you are publishing.

ifMatch :: Text

The current version (ETag value) of the function that you are publishing, which you can get using DescribeFunction.

newPublishFunction

Arguments

:: Text
$sel:name:PublishFunction'

-> Text
$sel:ifMatch:PublishFunction'

-> PublishFunction

Create a value of PublishFunction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:PublishFunction', publishFunction_name - The name of the function that you are publishing.

$sel:ifMatch:PublishFunction', publishFunction_ifMatch - The current version (ETag value) of the function that you are publishing, which you can get using DescribeFunction.

Request Lenses

publishFunction_name :: Lens' PublishFunction Text

The name of the function that you are publishing.

publishFunction_ifMatch :: Lens' PublishFunction Text

The current version (ETag value) of the function that you are publishing, which you can get using DescribeFunction.

Destructuring the Response

data PublishFunctionResponse

See: newPublishFunctionResponse smart constructor.

Constructors

PublishFunctionResponse'

Fields

functionSummary :: Maybe FunctionSummary

Contains configuration information and metadata about a CloudFront function.

httpStatus :: Int

The response's http status code.

newPublishFunctionResponse

Arguments

:: Int
$sel:httpStatus:PublishFunctionResponse'

-> PublishFunctionResponse

Create a value of PublishFunctionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:functionSummary:PublishFunctionResponse', publishFunctionResponse_functionSummary - Contains configuration information and metadata about a CloudFront function.

$sel:httpStatus:PublishFunctionResponse', publishFunctionResponse_httpStatus - The response's http status code.

Response Lenses

publishFunctionResponse_functionSummary :: Lens' PublishFunctionResponse (Maybe FunctionSummary)

Contains configuration information and metadata about a CloudFront function.

publishFunctionResponse_httpStatus :: Lens' PublishFunctionResponse Int

The response's http status code.