Network.AWS.CloudFront.TestFunction

Description

Tests a CloudFront function.

To test a function, you provide an event object that represents an HTTP request or response that your CloudFront distribution could receive in production. CloudFront runs the function, passing it the event object that you provided, and returns the function’s result (the modified event object) in the response. The response also contains function logs and error messages, if any exist. For more information about testing functions, see Testing functions in the Amazon CloudFront Developer Guide.

To test a function, you provide the function’s name and version (ETag value) along with the event object. To get the function’s name and version, you can use ListFunctions and DescribeFunction.

Creating a Request

data TestFunction

See: newTestFunction smart constructor.

Constructors

TestFunction'

Fields

stage :: Maybe FunctionStage

The stage of the function that you are testing, either DEVELOPMENT or LIVE.

name :: Text

The name of the function that you are testing.

ifMatch :: Text

The current version (ETag value) of the function that you are testing, which you can get using DescribeFunction.

eventObject :: Sensitive Base64

The event object to test the function with. For more information about the structure of the event object, see Testing functions in the Amazon CloudFront Developer Guide.

newTestFunction

Arguments

:: Text
$sel:name:TestFunction'

-> Text
$sel:ifMatch:TestFunction'

-> ByteString
$sel:eventObject:TestFunction'

-> TestFunction

Create a value of TestFunction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stage:TestFunction', testFunction_stage - The stage of the function that you are testing, either DEVELOPMENT or LIVE.

$sel:name:TestFunction', testFunction_name - The name of the function that you are testing.

$sel:ifMatch:TestFunction', testFunction_ifMatch - The current version (ETag value) of the function that you are testing, which you can get using DescribeFunction.

$sel:eventObject:TestFunction', testFunction_eventObject - The event object to test the function with. For more information about the structure of the event object, see Testing functions in the Amazon CloudFront Developer Guide.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

Request Lenses

testFunction_stage :: Lens' TestFunction (Maybe FunctionStage)

The stage of the function that you are testing, either DEVELOPMENT or LIVE.

testFunction_name :: Lens' TestFunction Text

The name of the function that you are testing.

testFunction_ifMatch :: Lens' TestFunction Text

The current version (ETag value) of the function that you are testing, which you can get using DescribeFunction.

testFunction_eventObject :: Lens' TestFunction ByteString

The event object to test the function with. For more information about the structure of the event object, see Testing functions in the Amazon CloudFront Developer Guide.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

Destructuring the Response

data TestFunctionResponse

See: newTestFunctionResponse smart constructor.

Constructors

TestFunctionResponse'

Fields

testResult :: Maybe TestResult

An object that represents the result of running the function with the provided event object.

httpStatus :: Int

The response's http status code.

newTestFunctionResponse

Arguments

:: Int
$sel:httpStatus:TestFunctionResponse'

-> TestFunctionResponse

Create a value of TestFunctionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:testResult:TestFunctionResponse', testFunctionResponse_testResult - An object that represents the result of running the function with the provided event object.

$sel:httpStatus:TestFunctionResponse', testFunctionResponse_httpStatus - The response's http status code.

Response Lenses

testFunctionResponse_testResult :: Lens' TestFunctionResponse (Maybe TestResult)

An object that represents the result of running the function with the provided event object.

testFunctionResponse_httpStatus :: Lens' TestFunctionResponse Int

The response's http status code.