Network.AWS.CloudFront.Types.CachePolicy

Description

Documentation

data CachePolicy

A cache policy.

When it’s attached to a cache behavior, the cache policy determines the following:

The values that CloudFront includes in the cache key. These values can include HTTP headers, cookies, and URL query strings. CloudFront uses the cache key to find an object in its cache that it can return to the viewer.

The default, minimum, and maximum time to live (TTL) values that you want objects to stay in the CloudFront cache.

The headers, cookies, and query strings that are included in the cache key are automatically included in requests that CloudFront sends to the origin. CloudFront sends a request when it can’t find a valid object in its cache that matches the request’s cache key. If you want to send values to the origin but not include them in the cache key, use OriginRequestPolicy.

See: newCachePolicy smart constructor.

Constructors

CachePolicy'

Fields

id :: Text

The unique identifier for the cache policy.

lastModifiedTime :: ISO8601

The date and time when the cache policy was last modified.

cachePolicyConfig :: CachePolicyConfig

The cache policy configuration.

newCachePolicy

Arguments

:: Text
$sel:id:CachePolicy'

-> UTCTime
$sel:lastModifiedTime:CachePolicy'

-> CachePolicyConfig
$sel:cachePolicyConfig:CachePolicy'

-> CachePolicy

Create a value of CachePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:CachePolicy', cachePolicy_id - The unique identifier for the cache policy.

$sel:lastModifiedTime:CachePolicy', cachePolicy_lastModifiedTime - The date and time when the cache policy was last modified.

$sel:cachePolicyConfig:CachePolicy', cachePolicy_cachePolicyConfig - The cache policy configuration.

cachePolicy_id :: Lens' CachePolicy Text

The unique identifier for the cache policy.

cachePolicy_lastModifiedTime :: Lens' CachePolicy UTCTime

The date and time when the cache policy was last modified.

cachePolicy_cachePolicyConfig :: Lens' CachePolicy CachePolicyConfig

The cache policy configuration.