Network.AWS.CloudFront.Types.CloudFrontOriginAccessIdentityConfig

Description

Documentation

data CloudFrontOriginAccessIdentityConfig

Origin access identity configuration. Send a GET request to the /CloudFront API version/CloudFront/identity ID/config resource.

See: newCloudFrontOriginAccessIdentityConfig smart constructor.

Constructors

CloudFrontOriginAccessIdentityConfig'

Fields

callerReference :: Text

A unique value (for example, a date-time stamp) that ensures that the request can't be replayed.

If the value of CallerReference is new (regardless of the content of the CloudFrontOriginAccessIdentityConfig object), a new origin access identity is created.

If the CallerReference is a value already sent in a previous identity request, and the content of the CloudFrontOriginAccessIdentityConfig is identical to the original request (ignoring white space), the response includes the same information returned to the original request.

If the CallerReference is a value you already sent in a previous request to create an identity, but the content of the CloudFrontOriginAccessIdentityConfig is different from the original request, CloudFront returns a CloudFrontOriginAccessIdentityAlreadyExists error.

comment :: Text

A comment to describe the origin access identity. The comment cannot be longer than 128 characters.

newCloudFrontOriginAccessIdentityConfig

Arguments

:: Text
$sel:callerReference:CloudFrontOriginAccessIdentityConfig'

-> Text
$sel:comment:CloudFrontOriginAccessIdentityConfig'

-> CloudFrontOriginAccessIdentityConfig

Create a value of CloudFrontOriginAccessIdentityConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:callerReference:CloudFrontOriginAccessIdentityConfig', cloudFrontOriginAccessIdentityConfig_callerReference - A unique value (for example, a date-time stamp) that ensures that the request can't be replayed.

If the value of CallerReference is new (regardless of the content of the CloudFrontOriginAccessIdentityConfig object), a new origin access identity is created.

If the CallerReference is a value already sent in a previous identity request, and the content of the CloudFrontOriginAccessIdentityConfig is identical to the original request (ignoring white space), the response includes the same information returned to the original request.

If the CallerReference is a value you already sent in a previous request to create an identity, but the content of the CloudFrontOriginAccessIdentityConfig is different from the original request, CloudFront returns a CloudFrontOriginAccessIdentityAlreadyExists error.

$sel:comment:CloudFrontOriginAccessIdentityConfig', cloudFrontOriginAccessIdentityConfig_comment - A comment to describe the origin access identity. The comment cannot be longer than 128 characters.

cloudFrontOriginAccessIdentityConfig_callerReference :: Lens' CloudFrontOriginAccessIdentityConfig Text

A unique value (for example, a date-time stamp) that ensures that the request can't be replayed.

If the value of CallerReference is new (regardless of the content of the CloudFrontOriginAccessIdentityConfig object), a new origin access identity is created.

If the CallerReference is a value already sent in a previous identity request, and the content of the CloudFrontOriginAccessIdentityConfig is identical to the original request (ignoring white space), the response includes the same information returned to the original request.

If the CallerReference is a value you already sent in a previous request to create an identity, but the content of the CloudFrontOriginAccessIdentityConfig is different from the original request, CloudFront returns a CloudFrontOriginAccessIdentityAlreadyExists error.

cloudFrontOriginAccessIdentityConfig_comment :: Lens' CloudFrontOriginAccessIdentityConfig Text

A comment to describe the origin access identity. The comment cannot be longer than 128 characters.