Network.AWS.CloudFront.Types.CloudFrontOriginAccessIdentityList

Description

Documentation

data CloudFrontOriginAccessIdentityList

Lists the origin access identities for CloudFront.Send a GET request to the /CloudFront API version/origin-access-identity/cloudfront resource. The response includes a CloudFrontOriginAccessIdentityList element with zero or more CloudFrontOriginAccessIdentitySummary child elements. By default, your entire list of origin access identities is returned in one single page. If the list is long, you can paginate it using the MaxItems and Marker parameters.

See: newCloudFrontOriginAccessIdentityList smart constructor.

Constructors

CloudFrontOriginAccessIdentityList'

Fields

items :: Maybe [CloudFrontOriginAccessIdentitySummary]

A complex type that contains one CloudFrontOriginAccessIdentitySummary element for each origin access identity that was created by the current account.

nextMarker :: Maybe Text

If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your origin access identities where they left off.

marker :: Text

Use this when paginating results to indicate where to begin in your list of origin access identities. The results include identities in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last identity on that page).

maxItems :: Int

The maximum number of origin access identities you want in the response body.

isTruncated :: Bool

A flag that indicates whether more origin access identities remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more items in the list.

quantity :: Int

The number of CloudFront origin access identities that were created by the current account.

newCloudFrontOriginAccessIdentityList

Arguments

:: Text
$sel:marker:CloudFrontOriginAccessIdentityList'

-> Int
$sel:maxItems:CloudFrontOriginAccessIdentityList'

-> Bool
$sel:isTruncated:CloudFrontOriginAccessIdentityList'

-> Int
$sel:quantity:CloudFrontOriginAccessIdentityList'

-> CloudFrontOriginAccessIdentityList

Create a value of CloudFrontOriginAccessIdentityList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:CloudFrontOriginAccessIdentityList', cloudFrontOriginAccessIdentityList_items - A complex type that contains one CloudFrontOriginAccessIdentitySummary element for each origin access identity that was created by the current account.

$sel:nextMarker:CloudFrontOriginAccessIdentityList', cloudFrontOriginAccessIdentityList_nextMarker - If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your origin access identities where they left off.

$sel:marker:CloudFrontOriginAccessIdentityList', cloudFrontOriginAccessIdentityList_marker - Use this when paginating results to indicate where to begin in your list of origin access identities. The results include identities in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last identity on that page).

$sel:maxItems:CloudFrontOriginAccessIdentityList', cloudFrontOriginAccessIdentityList_maxItems - The maximum number of origin access identities you want in the response body.

$sel:isTruncated:CloudFrontOriginAccessIdentityList', cloudFrontOriginAccessIdentityList_isTruncated - A flag that indicates whether more origin access identities remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more items in the list.

$sel:quantity:CloudFrontOriginAccessIdentityList', cloudFrontOriginAccessIdentityList_quantity - The number of CloudFront origin access identities that were created by the current account.

cloudFrontOriginAccessIdentityList_items :: Lens' CloudFrontOriginAccessIdentityList (Maybe [CloudFrontOriginAccessIdentitySummary])

A complex type that contains one CloudFrontOriginAccessIdentitySummary element for each origin access identity that was created by the current account.

cloudFrontOriginAccessIdentityList_nextMarker :: Lens' CloudFrontOriginAccessIdentityList (Maybe Text)

If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your origin access identities where they left off.

cloudFrontOriginAccessIdentityList_marker :: Lens' CloudFrontOriginAccessIdentityList Text

Use this when paginating results to indicate where to begin in your list of origin access identities. The results include identities in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last identity on that page).

cloudFrontOriginAccessIdentityList_maxItems :: Lens' CloudFrontOriginAccessIdentityList Int

The maximum number of origin access identities you want in the response body.

cloudFrontOriginAccessIdentityList_isTruncated :: Lens' CloudFrontOriginAccessIdentityList Bool

A flag that indicates whether more origin access identities remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more items in the list.

cloudFrontOriginAccessIdentityList_quantity :: Lens' CloudFrontOriginAccessIdentityList Int

The number of CloudFront origin access identities that were created by the current account.