Network.AWS.CloudFront.Types.CustomOriginConfig

Description

Documentation

data CustomOriginConfig

A custom origin. A custom origin is any origin that is not an Amazon S3 bucket, with one exception. An Amazon S3 bucket that is configured with static website hosting is a custom origin.

See: newCustomOriginConfig smart constructor.

Constructors

CustomOriginConfig'

Fields

originKeepaliveTimeout :: Maybe Int

Specifies how long, in seconds, CloudFront persists its connection to the origin. The minimum timeout is 1 second, the maximum is 60 seconds, and the default (if you don’t specify otherwise) is 5 seconds.

For more information, see Origin Keep-alive Timeout in the Amazon CloudFront Developer Guide.

originSslProtocols :: Maybe OriginSslProtocols

Specifies the minimum SSL/TLS protocol that CloudFront uses when connecting to your origin over HTTPS. Valid values include SSLv3, TLSv1, TLSv1.1, and TLSv1.2.

For more information, see Minimum Origin SSL Protocol in the Amazon CloudFront Developer Guide.

originReadTimeout :: Maybe Int

Specifies how long, in seconds, CloudFront waits for a response from the origin. This is also known as the origin response timeout. The minimum timeout is 1 second, the maximum is 60 seconds, and the default (if you don’t specify otherwise) is 30 seconds.

For more information, see Origin Response Timeout in the Amazon CloudFront Developer Guide.

hTTPPort :: Int

The HTTP port that CloudFront uses to connect to the origin. Specify the HTTP port that the origin listens on.

hTTPSPort :: Int

The HTTPS port that CloudFront uses to connect to the origin. Specify the HTTPS port that the origin listens on.

originProtocolPolicy :: OriginProtocolPolicy

Specifies the protocol (HTTP or HTTPS) that CloudFront uses to connect to the origin. Valid values are:

http-only – CloudFront always uses HTTP to connect to the origin.

match-viewer – CloudFront connects to the origin using the same protocol that the viewer used to connect to CloudFront.

https-only – CloudFront always uses HTTPS to connect to the origin.

newCustomOriginConfig

Arguments

:: Int
$sel:hTTPPort:CustomOriginConfig'

-> Int
$sel:hTTPSPort:CustomOriginConfig'

-> OriginProtocolPolicy
$sel:originProtocolPolicy:CustomOriginConfig'

-> CustomOriginConfig

Create a value of CustomOriginConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:originKeepaliveTimeout:CustomOriginConfig', customOriginConfig_originKeepaliveTimeout - Specifies how long, in seconds, CloudFront persists its connection to the origin. The minimum timeout is 1 second, the maximum is 60 seconds, and the default (if you don’t specify otherwise) is 5 seconds.

For more information, see Origin Keep-alive Timeout in the Amazon CloudFront Developer Guide.

$sel:originSslProtocols:CustomOriginConfig', customOriginConfig_originSslProtocols - Specifies the minimum SSL/TLS protocol that CloudFront uses when connecting to your origin over HTTPS. Valid values include SSLv3, TLSv1, TLSv1.1, and TLSv1.2.

For more information, see Minimum Origin SSL Protocol in the Amazon CloudFront Developer Guide.

$sel:originReadTimeout:CustomOriginConfig', customOriginConfig_originReadTimeout - Specifies how long, in seconds, CloudFront waits for a response from the origin. This is also known as the origin response timeout. The minimum timeout is 1 second, the maximum is 60 seconds, and the default (if you don’t specify otherwise) is 30 seconds.

For more information, see Origin Response Timeout in the Amazon CloudFront Developer Guide.

$sel:hTTPPort:CustomOriginConfig', customOriginConfig_hTTPPort - The HTTP port that CloudFront uses to connect to the origin. Specify the HTTP port that the origin listens on.

$sel:hTTPSPort:CustomOriginConfig', customOriginConfig_hTTPSPort - The HTTPS port that CloudFront uses to connect to the origin. Specify the HTTPS port that the origin listens on.

$sel:originProtocolPolicy:CustomOriginConfig', customOriginConfig_originProtocolPolicy - Specifies the protocol (HTTP or HTTPS) that CloudFront uses to connect to the origin. Valid values are:

http-only – CloudFront always uses HTTP to connect to the origin.

match-viewer – CloudFront connects to the origin using the same protocol that the viewer used to connect to CloudFront.

https-only – CloudFront always uses HTTPS to connect to the origin.

customOriginConfig_originKeepaliveTimeout :: Lens' CustomOriginConfig (Maybe Int)

Specifies how long, in seconds, CloudFront persists its connection to the origin. The minimum timeout is 1 second, the maximum is 60 seconds, and the default (if you don’t specify otherwise) is 5 seconds.

For more information, see Origin Keep-alive Timeout in the Amazon CloudFront Developer Guide.

customOriginConfig_originSslProtocols :: Lens' CustomOriginConfig (Maybe OriginSslProtocols)

Specifies the minimum SSL/TLS protocol that CloudFront uses when connecting to your origin over HTTPS. Valid values include SSLv3, TLSv1, TLSv1.1, and TLSv1.2.

For more information, see Minimum Origin SSL Protocol in the Amazon CloudFront Developer Guide.

customOriginConfig_originReadTimeout :: Lens' CustomOriginConfig (Maybe Int)

Specifies how long, in seconds, CloudFront waits for a response from the origin. This is also known as the origin response timeout. The minimum timeout is 1 second, the maximum is 60 seconds, and the default (if you don’t specify otherwise) is 30 seconds.

For more information, see Origin Response Timeout in the Amazon CloudFront Developer Guide.

customOriginConfig_hTTPPort :: Lens' CustomOriginConfig Int

The HTTP port that CloudFront uses to connect to the origin. Specify the HTTP port that the origin listens on.

customOriginConfig_hTTPSPort :: Lens' CustomOriginConfig Int

The HTTPS port that CloudFront uses to connect to the origin. Specify the HTTPS port that the origin listens on.

customOriginConfig_originProtocolPolicy :: Lens' CustomOriginConfig OriginProtocolPolicy

Specifies the protocol (HTTP or HTTPS) that CloudFront uses to connect to the origin. Valid values are:

http-only – CloudFront always uses HTTP to connect to the origin.

match-viewer – CloudFront connects to the origin using the same protocol that the viewer used to connect to CloudFront.

https-only – CloudFront always uses HTTPS to connect to the origin.