Network.AWS.CloudFront.Types.Distribution

Description

Documentation

data Distribution

A distribution tells CloudFront where you want content to be delivered from, and the details about how to track and manage content delivery.

See: newDistribution smart constructor.

Constructors

Distribution'

Fields

aliasICPRecordals :: Maybe [AliasICPRecordal]

Amazon Web Services services in China customers must file for an Internet Content Provider (ICP) recordal if they want to serve content publicly on an alternate domain name, also known as a CNAME, that they've added to CloudFront. AliasICPRecordal provides the ICP recordal status for CNAMEs associated with distributions.

For more information about ICP recordals, see Signup, Accounts, and Credentials in Getting Started with Amazon Web Services services in China.

activeTrustedSigners :: Maybe ActiveTrustedSigners

We recommend using TrustedKeyGroups instead of TrustedSigners.

CloudFront automatically adds this field to the response if you’ve configured a cache behavior in this distribution to serve private content using trusted signers. This field contains a list of account IDs and the active CloudFront key pairs in each account that CloudFront can use to verify the signatures of signed URLs or signed cookies.

activeTrustedKeyGroups :: Maybe ActiveTrustedKeyGroups

CloudFront automatically adds this field to the response if you’ve configured a cache behavior in this distribution to serve private content using key groups. This field contains a list of key groups and the public keys in each key group that CloudFront can use to verify the signatures of signed URLs or signed cookies.

id :: Text

The identifier for the distribution. For example: EDFDVBD632BHDS5.

arn :: Text

The ARN (Amazon Resource Name) for the distribution. For example: arn:aws:cloudfront::123456789012:distribution/EDFDVBD632BHDS5, where 123456789012 is your account ID.

status :: Text

This response element indicates the current status of the distribution. When the status is Deployed, the distribution's information is fully propagated to all CloudFront edge locations.

lastModifiedTime :: ISO8601

The date and time the distribution was last modified.

inProgressInvalidationBatches :: Int

The number of invalidation batches currently in progress.

domainName :: Text

The domain name corresponding to the distribution, for example, d111111abcdef8.cloudfront.net.

distributionConfig :: DistributionConfig

The current configuration information for the distribution. Send a GET request to the /CloudFront API version/distribution ID/config resource.

newDistribution

Arguments

:: Text
$sel:id:Distribution'

-> Text
$sel:arn:Distribution'

-> Text
$sel:status:Distribution'

-> UTCTime
$sel:lastModifiedTime:Distribution'

-> Int
$sel:inProgressInvalidationBatches:Distribution'

-> Text
$sel:domainName:Distribution'

-> DistributionConfig
$sel:distributionConfig:Distribution'

-> Distribution

Create a value of Distribution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aliasICPRecordals:Distribution', distribution_aliasICPRecordals - Amazon Web Services services in China customers must file for an Internet Content Provider (ICP) recordal if they want to serve content publicly on an alternate domain name, also known as a CNAME, that they've added to CloudFront. AliasICPRecordal provides the ICP recordal status for CNAMEs associated with distributions.

For more information about ICP recordals, see Signup, Accounts, and Credentials in Getting Started with Amazon Web Services services in China.

$sel:activeTrustedSigners:Distribution', distribution_activeTrustedSigners - We recommend using TrustedKeyGroups instead of TrustedSigners.

CloudFront automatically adds this field to the response if you’ve configured a cache behavior in this distribution to serve private content using trusted signers. This field contains a list of account IDs and the active CloudFront key pairs in each account that CloudFront can use to verify the signatures of signed URLs or signed cookies.

$sel:activeTrustedKeyGroups:Distribution', distribution_activeTrustedKeyGroups - CloudFront automatically adds this field to the response if you’ve configured a cache behavior in this distribution to serve private content using key groups. This field contains a list of key groups and the public keys in each key group that CloudFront can use to verify the signatures of signed URLs or signed cookies.

$sel:id:Distribution', distribution_id - The identifier for the distribution. For example: EDFDVBD632BHDS5.

$sel:arn:Distribution', distribution_arn - The ARN (Amazon Resource Name) for the distribution. For example: arn:aws:cloudfront::123456789012:distribution/EDFDVBD632BHDS5, where 123456789012 is your account ID.

$sel:status:Distribution', distribution_status - This response element indicates the current status of the distribution. When the status is Deployed, the distribution's information is fully propagated to all CloudFront edge locations.

$sel:lastModifiedTime:Distribution', distribution_lastModifiedTime - The date and time the distribution was last modified.

$sel:inProgressInvalidationBatches:Distribution', distribution_inProgressInvalidationBatches - The number of invalidation batches currently in progress.

$sel:domainName:Distribution', distribution_domainName - The domain name corresponding to the distribution, for example, d111111abcdef8.cloudfront.net.

$sel:distributionConfig:Distribution', distribution_distributionConfig - The current configuration information for the distribution. Send a GET request to the /CloudFront API version/distribution ID/config resource.

distribution_aliasICPRecordals :: Lens' Distribution (Maybe [AliasICPRecordal])

Amazon Web Services services in China customers must file for an Internet Content Provider (ICP) recordal if they want to serve content publicly on an alternate domain name, also known as a CNAME, that they've added to CloudFront. AliasICPRecordal provides the ICP recordal status for CNAMEs associated with distributions.

For more information about ICP recordals, see Signup, Accounts, and Credentials in Getting Started with Amazon Web Services services in China.

distribution_activeTrustedSigners :: Lens' Distribution (Maybe ActiveTrustedSigners)

We recommend using TrustedKeyGroups instead of TrustedSigners.

CloudFront automatically adds this field to the response if you’ve configured a cache behavior in this distribution to serve private content using trusted signers. This field contains a list of account IDs and the active CloudFront key pairs in each account that CloudFront can use to verify the signatures of signed URLs or signed cookies.

distribution_activeTrustedKeyGroups :: Lens' Distribution (Maybe ActiveTrustedKeyGroups)

CloudFront automatically adds this field to the response if you’ve configured a cache behavior in this distribution to serve private content using key groups. This field contains a list of key groups and the public keys in each key group that CloudFront can use to verify the signatures of signed URLs or signed cookies.

distribution_id :: Lens' Distribution Text

The identifier for the distribution. For example: EDFDVBD632BHDS5.

distribution_arn :: Lens' Distribution Text

The ARN (Amazon Resource Name) for the distribution. For example: arn:aws:cloudfront::123456789012:distribution/EDFDVBD632BHDS5, where 123456789012 is your account ID.

distribution_status :: Lens' Distribution Text

This response element indicates the current status of the distribution. When the status is Deployed, the distribution's information is fully propagated to all CloudFront edge locations.

distribution_lastModifiedTime :: Lens' Distribution UTCTime

The date and time the distribution was last modified.

distribution_inProgressInvalidationBatches :: Lens' Distribution Int

The number of invalidation batches currently in progress.

distribution_domainName :: Lens' Distribution Text

The domain name corresponding to the distribution, for example, d111111abcdef8.cloudfront.net.

distribution_distributionConfig :: Lens' Distribution DistributionConfig

The current configuration information for the distribution. Send a GET request to the /CloudFront API version/distribution ID/config resource.