Network.AWS.CloudFront.Types.DistributionIdList

Description

Documentation

data DistributionIdList

A list of distribution IDs.

See: newDistributionIdList smart constructor.

Constructors

DistributionIdList'

Fields

items :: Maybe [Text]

Contains the distribution IDs in the list.

nextMarker :: Maybe Text

Contains the value that you should use in the Marker field of a subsequent request to continue listing distribution IDs where you left off.

marker :: Text

The value provided in the Marker request field.

maxItems :: Int

The maximum number of distribution IDs requested.

isTruncated :: Bool

A flag that indicates whether more distribution IDs remain to be listed. If your results were truncated, you can make a subsequent request using the Marker request field to retrieve more distribution IDs in the list.

quantity :: Int

The total number of distribution IDs returned in the response.

newDistributionIdList

Arguments

:: Text
$sel:marker:DistributionIdList'

-> Int
$sel:maxItems:DistributionIdList'

-> Bool
$sel:isTruncated:DistributionIdList'

-> Int
$sel:quantity:DistributionIdList'

-> DistributionIdList

Create a value of DistributionIdList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:DistributionIdList', distributionIdList_items - Contains the distribution IDs in the list.

$sel:nextMarker:DistributionIdList', distributionIdList_nextMarker - Contains the value that you should use in the Marker field of a subsequent request to continue listing distribution IDs where you left off.

$sel:marker:DistributionIdList', distributionIdList_marker - The value provided in the Marker request field.

$sel:maxItems:DistributionIdList', distributionIdList_maxItems - The maximum number of distribution IDs requested.

$sel:isTruncated:DistributionIdList', distributionIdList_isTruncated - A flag that indicates whether more distribution IDs remain to be listed. If your results were truncated, you can make a subsequent request using the Marker request field to retrieve more distribution IDs in the list.

$sel:quantity:DistributionIdList', distributionIdList_quantity - The total number of distribution IDs returned in the response.

distributionIdList_items :: Lens' DistributionIdList (Maybe [Text])

Contains the distribution IDs in the list.

distributionIdList_nextMarker :: Lens' DistributionIdList (Maybe Text)

Contains the value that you should use in the Marker field of a subsequent request to continue listing distribution IDs where you left off.

distributionIdList_marker :: Lens' DistributionIdList Text

The value provided in the Marker request field.

distributionIdList_maxItems :: Lens' DistributionIdList Int

The maximum number of distribution IDs requested.

distributionIdList_isTruncated :: Lens' DistributionIdList Bool

A flag that indicates whether more distribution IDs remain to be listed. If your results were truncated, you can make a subsequent request using the Marker request field to retrieve more distribution IDs in the list.

distributionIdList_quantity :: Lens' DistributionIdList Int

The total number of distribution IDs returned in the response.