Network.AWS.CloudFront.Types.DistributionList

Description

Documentation

data DistributionList

A distribution list.

See: newDistributionList smart constructor.

Constructors

DistributionList'

Fields

items :: Maybe [DistributionSummary]

A complex type that contains one DistributionSummary element for each distribution that was created by the current account.

nextMarker :: Maybe Text

If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your distributions where they left off.

marker :: Text

The value you provided for the Marker request parameter.

maxItems :: Int

The value you provided for the MaxItems request parameter.

isTruncated :: Bool

A flag that indicates whether more distributions remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more distributions in the list.

quantity :: Int

The number of distributions that were created by the current account.

newDistributionList

Arguments

:: Text
$sel:marker:DistributionList'

-> Int
$sel:maxItems:DistributionList'

-> Bool
$sel:isTruncated:DistributionList'

-> Int
$sel:quantity:DistributionList'

-> DistributionList

Create a value of DistributionList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:DistributionList', distributionList_items - A complex type that contains one DistributionSummary element for each distribution that was created by the current account.

$sel:nextMarker:DistributionList', distributionList_nextMarker - If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your distributions where they left off.

$sel:marker:DistributionList', distributionList_marker - The value you provided for the Marker request parameter.

$sel:maxItems:DistributionList', distributionList_maxItems - The value you provided for the MaxItems request parameter.

$sel:isTruncated:DistributionList', distributionList_isTruncated - A flag that indicates whether more distributions remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more distributions in the list.

$sel:quantity:DistributionList', distributionList_quantity - The number of distributions that were created by the current account.

distributionList_items :: Lens' DistributionList (Maybe [DistributionSummary])

A complex type that contains one DistributionSummary element for each distribution that was created by the current account.

distributionList_nextMarker :: Lens' DistributionList (Maybe Text)

If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your distributions where they left off.

distributionList_marker :: Lens' DistributionList Text

The value you provided for the Marker request parameter.

distributionList_maxItems :: Lens' DistributionList Int

The value you provided for the MaxItems request parameter.

distributionList_isTruncated :: Lens' DistributionList Bool

A flag that indicates whether more distributions remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more distributions in the list.

distributionList_quantity :: Lens' DistributionList Int

The number of distributions that were created by the current account.