Network.AWS.CloudFront.Types.DistributionSummary

Description

Documentation

data DistributionSummary

A summary of the information about a CloudFront distribution.

See: newDistributionSummary smart constructor.

Constructors

DistributionSummary'

Fields

aliasICPRecordals :: Maybe [AliasICPRecordal]

Amazon Web Services services in China customers must file for an Internet Content Provider (ICP) recordal if they want to serve content publicly on an alternate domain name, also known as a CNAME, that they've added to CloudFront. AliasICPRecordal provides the ICP recordal status for CNAMEs associated with distributions.

For more information about ICP recordals, see Signup, Accounts, and Credentials in Getting Started with Amazon Web Services services in China.

originGroups :: Maybe OriginGroups

A complex type that contains information about origin groups for this distribution.

id :: Text

The identifier for the distribution. For example: EDFDVBD632BHDS5.

arn :: Text

The ARN (Amazon Resource Name) for the distribution. For example: arn:aws:cloudfront::123456789012:distribution/EDFDVBD632BHDS5, where 123456789012 is your account ID.

status :: Text

The current status of the distribution. When the status is Deployed, the distribution's information is propagated to all CloudFront edge locations.

lastModifiedTime :: ISO8601

The date and time the distribution was last modified.

domainName :: Text

The domain name that corresponds to the distribution, for example, d111111abcdef8.cloudfront.net.

aliases :: Aliases

A complex type that contains information about CNAMEs (alternate domain names), if any, for this distribution.

origins :: Origins

A complex type that contains information about origins for this distribution.

defaultCacheBehavior :: DefaultCacheBehavior

A complex type that describes the default cache behavior if you don't specify a CacheBehavior element or if files don't match any of the values of PathPattern in CacheBehavior elements. You must create exactly one default cache behavior.

cacheBehaviors :: CacheBehaviors

A complex type that contains zero or more CacheBehavior elements.

customErrorResponses :: CustomErrorResponses

A complex type that contains zero or more CustomErrorResponses elements.

comment :: Text

The comment originally specified when this distribution was created.

priceClass :: PriceClass

A complex type that contains information about price class for this streaming distribution.

enabled :: Bool

Whether the distribution is enabled to accept user requests for content.

viewerCertificate :: ViewerCertificate

A complex type that determines the distribution’s SSL/TLS configuration for communicating with viewers.

restrictions :: Restrictions

A complex type that identifies ways in which you want to restrict distribution of your content.

webACLId :: Text

The Web ACL Id (if any) associated with the distribution.

httpVersion :: HttpVersion

Specify the maximum HTTP version that you want viewers to use to communicate with CloudFront. The default value for new web distributions is http2. Viewers that don't support HTTP/2 will automatically use an earlier version.

isIPV6Enabled :: Bool

Whether CloudFront responds to IPv6 DNS requests with an IPv6 address for your distribution.

newDistributionSummary

Arguments

:: Text
$sel:id:DistributionSummary'

-> Text
$sel:arn:DistributionSummary'

-> Text
$sel:status:DistributionSummary'

-> UTCTime
$sel:lastModifiedTime:DistributionSummary'

-> Text
$sel:domainName:DistributionSummary'

-> Aliases
$sel:aliases:DistributionSummary'

-> Origins
$sel:origins:DistributionSummary'

-> DefaultCacheBehavior
$sel:defaultCacheBehavior:DistributionSummary'

-> CacheBehaviors
$sel:cacheBehaviors:DistributionSummary'

-> CustomErrorResponses
$sel:customErrorResponses:DistributionSummary'

-> Text
$sel:comment:DistributionSummary'

-> PriceClass
$sel:priceClass:DistributionSummary'

-> Bool
$sel:enabled:DistributionSummary'

-> ViewerCertificate
$sel:viewerCertificate:DistributionSummary'

-> Restrictions
$sel:restrictions:DistributionSummary'

-> Text
$sel:webACLId:DistributionSummary'

-> HttpVersion
$sel:httpVersion:DistributionSummary'

-> Bool
$sel:isIPV6Enabled:DistributionSummary'

-> DistributionSummary

Create a value of DistributionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aliasICPRecordals:DistributionSummary', distributionSummary_aliasICPRecordals - Amazon Web Services services in China customers must file for an Internet Content Provider (ICP) recordal if they want to serve content publicly on an alternate domain name, also known as a CNAME, that they've added to CloudFront. AliasICPRecordal provides the ICP recordal status for CNAMEs associated with distributions.

For more information about ICP recordals, see Signup, Accounts, and Credentials in Getting Started with Amazon Web Services services in China.

$sel:originGroups:DistributionSummary', distributionSummary_originGroups - A complex type that contains information about origin groups for this distribution.

$sel:id:DistributionSummary', distributionSummary_id - The identifier for the distribution. For example: EDFDVBD632BHDS5.

$sel:arn:DistributionSummary', distributionSummary_arn - The ARN (Amazon Resource Name) for the distribution. For example: arn:aws:cloudfront::123456789012:distribution/EDFDVBD632BHDS5, where 123456789012 is your account ID.

$sel:status:DistributionSummary', distributionSummary_status - The current status of the distribution. When the status is Deployed, the distribution's information is propagated to all CloudFront edge locations.

$sel:lastModifiedTime:DistributionSummary', distributionSummary_lastModifiedTime - The date and time the distribution was last modified.

$sel:domainName:DistributionSummary', distributionSummary_domainName - The domain name that corresponds to the distribution, for example, d111111abcdef8.cloudfront.net.

$sel:aliases:DistributionSummary', distributionSummary_aliases - A complex type that contains information about CNAMEs (alternate domain names), if any, for this distribution.

$sel:origins:DistributionSummary', distributionSummary_origins - A complex type that contains information about origins for this distribution.

$sel:defaultCacheBehavior:DistributionSummary', distributionSummary_defaultCacheBehavior - A complex type that describes the default cache behavior if you don't specify a CacheBehavior element or if files don't match any of the values of PathPattern in CacheBehavior elements. You must create exactly one default cache behavior.

$sel:cacheBehaviors:DistributionSummary', distributionSummary_cacheBehaviors - A complex type that contains zero or more CacheBehavior elements.

$sel:customErrorResponses:DistributionSummary', distributionSummary_customErrorResponses - A complex type that contains zero or more CustomErrorResponses elements.

$sel:comment:DistributionSummary', distributionSummary_comment - The comment originally specified when this distribution was created.

$sel:priceClass:DistributionSummary', distributionSummary_priceClass - A complex type that contains information about price class for this streaming distribution.

$sel:enabled:DistributionSummary', distributionSummary_enabled - Whether the distribution is enabled to accept user requests for content.

$sel:viewerCertificate:DistributionSummary', distributionSummary_viewerCertificate - A complex type that determines the distribution’s SSL/TLS configuration for communicating with viewers.

$sel:restrictions:DistributionSummary', distributionSummary_restrictions - A complex type that identifies ways in which you want to restrict distribution of your content.

$sel:webACLId:DistributionSummary', distributionSummary_webACLId - The Web ACL Id (if any) associated with the distribution.

$sel:httpVersion:DistributionSummary', distributionSummary_httpVersion - Specify the maximum HTTP version that you want viewers to use to communicate with CloudFront. The default value for new web distributions is http2. Viewers that don't support HTTP/2 will automatically use an earlier version.

$sel:isIPV6Enabled:DistributionSummary', distributionSummary_isIPV6Enabled - Whether CloudFront responds to IPv6 DNS requests with an IPv6 address for your distribution.

distributionSummary_aliasICPRecordals :: Lens' DistributionSummary (Maybe [AliasICPRecordal])

Amazon Web Services services in China customers must file for an Internet Content Provider (ICP) recordal if they want to serve content publicly on an alternate domain name, also known as a CNAME, that they've added to CloudFront. AliasICPRecordal provides the ICP recordal status for CNAMEs associated with distributions.

For more information about ICP recordals, see Signup, Accounts, and Credentials in Getting Started with Amazon Web Services services in China.

distributionSummary_originGroups :: Lens' DistributionSummary (Maybe OriginGroups)

A complex type that contains information about origin groups for this distribution.

distributionSummary_id :: Lens' DistributionSummary Text

The identifier for the distribution. For example: EDFDVBD632BHDS5.

distributionSummary_arn :: Lens' DistributionSummary Text

The ARN (Amazon Resource Name) for the distribution. For example: arn:aws:cloudfront::123456789012:distribution/EDFDVBD632BHDS5, where 123456789012 is your account ID.

distributionSummary_status :: Lens' DistributionSummary Text

The current status of the distribution. When the status is Deployed, the distribution's information is propagated to all CloudFront edge locations.

distributionSummary_lastModifiedTime :: Lens' DistributionSummary UTCTime

The date and time the distribution was last modified.

distributionSummary_domainName :: Lens' DistributionSummary Text

The domain name that corresponds to the distribution, for example, d111111abcdef8.cloudfront.net.

distributionSummary_aliases :: Lens' DistributionSummary Aliases

A complex type that contains information about CNAMEs (alternate domain names), if any, for this distribution.

distributionSummary_origins :: Lens' DistributionSummary Origins

A complex type that contains information about origins for this distribution.

distributionSummary_defaultCacheBehavior :: Lens' DistributionSummary DefaultCacheBehavior

A complex type that describes the default cache behavior if you don't specify a CacheBehavior element or if files don't match any of the values of PathPattern in CacheBehavior elements. You must create exactly one default cache behavior.

distributionSummary_cacheBehaviors :: Lens' DistributionSummary CacheBehaviors

A complex type that contains zero or more CacheBehavior elements.

distributionSummary_customErrorResponses :: Lens' DistributionSummary CustomErrorResponses

A complex type that contains zero or more CustomErrorResponses elements.

distributionSummary_comment :: Lens' DistributionSummary Text

The comment originally specified when this distribution was created.

distributionSummary_priceClass :: Lens' DistributionSummary PriceClass

A complex type that contains information about price class for this streaming distribution.

distributionSummary_enabled :: Lens' DistributionSummary Bool

Whether the distribution is enabled to accept user requests for content.

distributionSummary_viewerCertificate :: Lens' DistributionSummary ViewerCertificate

A complex type that determines the distribution’s SSL/TLS configuration for communicating with viewers.

distributionSummary_restrictions :: Lens' DistributionSummary Restrictions

A complex type that identifies ways in which you want to restrict distribution of your content.

distributionSummary_webACLId :: Lens' DistributionSummary Text

The Web ACL Id (if any) associated with the distribution.

distributionSummary_httpVersion :: Lens' DistributionSummary HttpVersion

Specify the maximum HTTP version that you want viewers to use to communicate with CloudFront. The default value for new web distributions is http2. Viewers that don't support HTTP/2 will automatically use an earlier version.

distributionSummary_isIPV6Enabled :: Lens' DistributionSummary Bool

Whether CloudFront responds to IPv6 DNS requests with an IPv6 address for your distribution.