Network.AWS.CloudFront.Types.Invalidation

Description

Documentation

data Invalidation

An invalidation.

See: newInvalidation smart constructor.

Constructors

Invalidation'

Fields

id :: Text

The identifier for the invalidation request. For example: IDFDVBD632BHDS5.

status :: Text

The status of the invalidation request. When the invalidation batch is finished, the status is Completed.

createTime :: ISO8601

The date and time the invalidation request was first made.

invalidationBatch :: InvalidationBatch

The current invalidation information for the batch request.

newInvalidation

Arguments

:: Text
$sel:id:Invalidation'

-> Text
$sel:status:Invalidation'

-> UTCTime
$sel:createTime:Invalidation'

-> InvalidationBatch
$sel:invalidationBatch:Invalidation'

-> Invalidation

Create a value of Invalidation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Invalidation', invalidation_id - The identifier for the invalidation request. For example: IDFDVBD632BHDS5.

$sel:status:Invalidation', invalidation_status - The status of the invalidation request. When the invalidation batch is finished, the status is Completed.

$sel:createTime:Invalidation', invalidation_createTime - The date and time the invalidation request was first made.

$sel:invalidationBatch:Invalidation', invalidation_invalidationBatch - The current invalidation information for the batch request.

invalidation_id :: Lens' Invalidation Text

The identifier for the invalidation request. For example: IDFDVBD632BHDS5.

invalidation_status :: Lens' Invalidation Text

The status of the invalidation request. When the invalidation batch is finished, the status is Completed.

invalidation_createTime :: Lens' Invalidation UTCTime

The date and time the invalidation request was first made.

invalidation_invalidationBatch :: Lens' Invalidation InvalidationBatch

The current invalidation information for the batch request.