Network.AWS.CloudFront.Types.KeyGroup

Description

Documentation

data KeyGroup

A key group.

A key group contains a list of public keys that you can use with CloudFront signed URLs and signed cookies.

See: newKeyGroup smart constructor.

Constructors

KeyGroup'

Fields

id :: Text

The identifier for the key group.

lastModifiedTime :: ISO8601

The date and time when the key group was last modified.

keyGroupConfig :: KeyGroupConfig

The key group configuration.

newKeyGroup

Arguments

:: Text
$sel:id:KeyGroup'

-> UTCTime
$sel:lastModifiedTime:KeyGroup'

-> KeyGroupConfig
$sel:keyGroupConfig:KeyGroup'

-> KeyGroup

Create a value of KeyGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:KeyGroup', keyGroup_id - The identifier for the key group.

$sel:lastModifiedTime:KeyGroup', keyGroup_lastModifiedTime - The date and time when the key group was last modified.

$sel:keyGroupConfig:KeyGroup', keyGroup_keyGroupConfig - The key group configuration.

keyGroup_id :: Lens' KeyGroup Text

The identifier for the key group.

keyGroup_lastModifiedTime :: Lens' KeyGroup UTCTime

The date and time when the key group was last modified.

keyGroup_keyGroupConfig :: Lens' KeyGroup KeyGroupConfig

The key group configuration.