Network.AWS.CloudFront.Types.LambdaFunctionAssociation

Description

Documentation

data LambdaFunctionAssociation

A complex type that contains a Lambda@Edge function association.

See: newLambdaFunctionAssociation smart constructor.

Constructors

LambdaFunctionAssociation'

Fields

includeBody :: Maybe Bool

A flag that allows a Lambda@Edge function to have read access to the body content. For more information, see Accessing the Request Body by Choosing the Include Body Option in the Amazon CloudFront Developer Guide.

lambdaFunctionARN :: Text

The ARN of the Lambda@Edge function. You must specify the ARN of a function version; you can't specify an alias or $LATEST.

eventType :: EventType

Specifies the event type that triggers a Lambda@Edge function invocation. You can specify the following values:

viewer-request: The function executes when CloudFront receives a request from a viewer and before it checks to see whether the requested object is in the edge cache.

origin-request: The function executes only when CloudFront sends a request to your origin. When the requested object is in the edge cache, the function doesn't execute.

origin-response: The function executes after CloudFront receives a response from the origin and before it caches the object in the response. When the requested object is in the edge cache, the function doesn't execute.

viewer-response: The function executes before CloudFront returns the requested object to the viewer. The function executes regardless of whether the object was already in the edge cache.

If the origin returns an HTTP status code other than HTTP 200 (OK), the function doesn't execute.

newLambdaFunctionAssociation

Arguments

:: Text
$sel:lambdaFunctionARN:LambdaFunctionAssociation'

-> EventType
$sel:eventType:LambdaFunctionAssociation'

-> LambdaFunctionAssociation

Create a value of LambdaFunctionAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeBody:LambdaFunctionAssociation', lambdaFunctionAssociation_includeBody - A flag that allows a Lambda@Edge function to have read access to the body content. For more information, see Accessing the Request Body by Choosing the Include Body Option in the Amazon CloudFront Developer Guide.

$sel:lambdaFunctionARN:LambdaFunctionAssociation', lambdaFunctionAssociation_lambdaFunctionARN - The ARN of the Lambda@Edge function. You must specify the ARN of a function version; you can't specify an alias or $LATEST.

$sel:eventType:LambdaFunctionAssociation', lambdaFunctionAssociation_eventType - Specifies the event type that triggers a Lambda@Edge function invocation. You can specify the following values:

viewer-request: The function executes when CloudFront receives a request from a viewer and before it checks to see whether the requested object is in the edge cache.

origin-request: The function executes only when CloudFront sends a request to your origin. When the requested object is in the edge cache, the function doesn't execute.

origin-response: The function executes after CloudFront receives a response from the origin and before it caches the object in the response. When the requested object is in the edge cache, the function doesn't execute.

viewer-response: The function executes before CloudFront returns the requested object to the viewer. The function executes regardless of whether the object was already in the edge cache.

If the origin returns an HTTP status code other than HTTP 200 (OK), the function doesn't execute.

lambdaFunctionAssociation_includeBody :: Lens' LambdaFunctionAssociation (Maybe Bool)

A flag that allows a Lambda@Edge function to have read access to the body content. For more information, see Accessing the Request Body by Choosing the Include Body Option in the Amazon CloudFront Developer Guide.

lambdaFunctionAssociation_lambdaFunctionARN :: Lens' LambdaFunctionAssociation Text

The ARN of the Lambda@Edge function. You must specify the ARN of a function version; you can't specify an alias or $LATEST.

lambdaFunctionAssociation_eventType :: Lens' LambdaFunctionAssociation EventType

Specifies the event type that triggers a Lambda@Edge function invocation. You can specify the following values:

viewer-request: The function executes when CloudFront receives a request from a viewer and before it checks to see whether the requested object is in the edge cache.

origin-request: The function executes only when CloudFront sends a request to your origin. When the requested object is in the edge cache, the function doesn't execute.

origin-response: The function executes after CloudFront receives a response from the origin and before it caches the object in the response. When the requested object is in the edge cache, the function doesn't execute.

viewer-response: The function executes before CloudFront returns the requested object to the viewer. The function executes regardless of whether the object was already in the edge cache.

If the origin returns an HTTP status code other than HTTP 200 (OK), the function doesn't execute.