Network.AWS.CloudFront.Types.OriginGroupFailoverCriteria

Description

Documentation

data OriginGroupFailoverCriteria

A complex data type that includes information about the failover criteria for an origin group, including the status codes for which CloudFront will failover from the primary origin to the second origin.

See: newOriginGroupFailoverCriteria smart constructor.

Constructors

OriginGroupFailoverCriteria'

Fields

statusCodes :: StatusCodes

The status codes that, when returned from the primary origin, will trigger CloudFront to failover to the second origin.

newOriginGroupFailoverCriteria

Arguments

:: StatusCodes
$sel:statusCodes:OriginGroupFailoverCriteria'

-> OriginGroupFailoverCriteria

Create a value of OriginGroupFailoverCriteria with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusCodes:OriginGroupFailoverCriteria', originGroupFailoverCriteria_statusCodes - The status codes that, when returned from the primary origin, will trigger CloudFront to failover to the second origin.

originGroupFailoverCriteria_statusCodes :: Lens' OriginGroupFailoverCriteria StatusCodes

The status codes that, when returned from the primary origin, will trigger CloudFront to failover to the second origin.