Network.AWS.CloudFront.Types.OriginRequestPolicyList

Description

Documentation

data OriginRequestPolicyList

A list of origin request policies.

See: newOriginRequestPolicyList smart constructor.

Constructors

OriginRequestPolicyList'

Fields

items :: Maybe [OriginRequestPolicySummary]

Contains the origin request policies in the list.

nextMarker :: Maybe Text

If there are more items in the list than are in this response, this element is present. It contains the value that you should use in the Marker field of a subsequent request to continue listing origin request policies where you left off.

maxItems :: Int

The maximum number of origin request policies requested.

quantity :: Int

The total number of origin request policies returned in the response.

newOriginRequestPolicyList

Arguments

:: Int
$sel:maxItems:OriginRequestPolicyList'

-> Int
$sel:quantity:OriginRequestPolicyList'

-> OriginRequestPolicyList

Create a value of OriginRequestPolicyList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:OriginRequestPolicyList', originRequestPolicyList_items - Contains the origin request policies in the list.

$sel:nextMarker:OriginRequestPolicyList', originRequestPolicyList_nextMarker - If there are more items in the list than are in this response, this element is present. It contains the value that you should use in the Marker field of a subsequent request to continue listing origin request policies where you left off.

$sel:maxItems:OriginRequestPolicyList', originRequestPolicyList_maxItems - The maximum number of origin request policies requested.

$sel:quantity:OriginRequestPolicyList', originRequestPolicyList_quantity - The total number of origin request policies returned in the response.

originRequestPolicyList_items :: Lens' OriginRequestPolicyList (Maybe [OriginRequestPolicySummary])

Contains the origin request policies in the list.

originRequestPolicyList_nextMarker :: Lens' OriginRequestPolicyList (Maybe Text)

If there are more items in the list than are in this response, this element is present. It contains the value that you should use in the Marker field of a subsequent request to continue listing origin request policies where you left off.

originRequestPolicyList_maxItems :: Lens' OriginRequestPolicyList Int

The maximum number of origin request policies requested.

originRequestPolicyList_quantity :: Lens' OriginRequestPolicyList Int

The total number of origin request policies returned in the response.