Network.AWS.CloudFront.Types.OriginRequestPolicyQueryStringsConfig

Description

Documentation

data OriginRequestPolicyQueryStringsConfig

An object that determines whether any URL query strings in viewer requests (and if so, which query strings) are included in requests that CloudFront sends to the origin.

See: newOriginRequestPolicyQueryStringsConfig smart constructor.

Constructors

OriginRequestPolicyQueryStringsConfig'

Fields

queryStrings :: Maybe QueryStringNames

Contains a list of the query strings in viewer requests that are included in requests that CloudFront sends to the origin.

queryStringBehavior :: OriginRequestPolicyQueryStringBehavior

Determines whether any URL query strings in viewer requests are included in requests that CloudFront sends to the origin. Valid values are:

none – Query strings in viewer requests are not included in requests that CloudFront sends to the origin. Even when this field is set to none, any query strings that are listed in a CachePolicy are included in origin requests.

whitelist – The query strings in viewer requests that are listed in the QueryStringNames type are included in requests that CloudFront sends to the origin.

all – All query strings in viewer requests are included in requests that CloudFront sends to the origin.

newOriginRequestPolicyQueryStringsConfig

Arguments

:: OriginRequestPolicyQueryStringBehavior
$sel:queryStringBehavior:OriginRequestPolicyQueryStringsConfig'

-> OriginRequestPolicyQueryStringsConfig

Create a value of OriginRequestPolicyQueryStringsConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queryStrings:OriginRequestPolicyQueryStringsConfig', originRequestPolicyQueryStringsConfig_queryStrings - Contains a list of the query strings in viewer requests that are included in requests that CloudFront sends to the origin.

$sel:queryStringBehavior:OriginRequestPolicyQueryStringsConfig', originRequestPolicyQueryStringsConfig_queryStringBehavior - Determines whether any URL query strings in viewer requests are included in requests that CloudFront sends to the origin. Valid values are:

none – Query strings in viewer requests are not included in requests that CloudFront sends to the origin. Even when this field is set to none, any query strings that are listed in a CachePolicy are included in origin requests.

whitelist – The query strings in viewer requests that are listed in the QueryStringNames type are included in requests that CloudFront sends to the origin.

all – All query strings in viewer requests are included in requests that CloudFront sends to the origin.

originRequestPolicyQueryStringsConfig_queryStrings :: Lens' OriginRequestPolicyQueryStringsConfig (Maybe QueryStringNames)

Contains a list of the query strings in viewer requests that are included in requests that CloudFront sends to the origin.

originRequestPolicyQueryStringsConfig_queryStringBehavior :: Lens' OriginRequestPolicyQueryStringsConfig OriginRequestPolicyQueryStringBehavior

Determines whether any URL query strings in viewer requests are included in requests that CloudFront sends to the origin. Valid values are:

none – Query strings in viewer requests are not included in requests that CloudFront sends to the origin. Even when this field is set to none, any query strings that are listed in a CachePolicy are included in origin requests.

whitelist – The query strings in viewer requests that are listed in the QueryStringNames type are included in requests that CloudFront sends to the origin.

all – All query strings in viewer requests are included in requests that CloudFront sends to the origin.