Network.AWS.CloudFront.Types.PublicKeySummary

Description

Documentation

data PublicKeySummary

Contains information about a public key.

See: newPublicKeySummary smart constructor.

Constructors

PublicKeySummary'

Fields

comment :: Maybe Text

A comment to describe the public key. The comment cannot be longer than 128 characters.

id :: Text

The identifier of the public key.

name :: Text

A name to help identify the public key.

createdTime :: ISO8601

The date and time when the public key was uploaded.

encodedKey :: Text

The public key.

newPublicKeySummary

Arguments

:: Text
$sel:id:PublicKeySummary'

-> Text
$sel:name:PublicKeySummary'

-> UTCTime
$sel:createdTime:PublicKeySummary'

-> Text
$sel:encodedKey:PublicKeySummary'

-> PublicKeySummary

Create a value of PublicKeySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comment:PublicKeySummary', publicKeySummary_comment - A comment to describe the public key. The comment cannot be longer than 128 characters.

$sel:id:PublicKeySummary', publicKeySummary_id - The identifier of the public key.

$sel:name:PublicKeySummary', publicKeySummary_name - A name to help identify the public key.

$sel:createdTime:PublicKeySummary', publicKeySummary_createdTime - The date and time when the public key was uploaded.

$sel:encodedKey:PublicKeySummary', publicKeySummary_encodedKey - The public key.

publicKeySummary_comment :: Lens' PublicKeySummary (Maybe Text)

A comment to describe the public key. The comment cannot be longer than 128 characters.

publicKeySummary_id :: Lens' PublicKeySummary Text

The identifier of the public key.

publicKeySummary_name :: Lens' PublicKeySummary Text

A name to help identify the public key.

publicKeySummary_createdTime :: Lens' PublicKeySummary UTCTime

The date and time when the public key was uploaded.

publicKeySummary_encodedKey :: Lens' PublicKeySummary Text

The public key.