Network.AWS.CloudFront.Types.StatusCodes

Description

Documentation

data StatusCodes

A complex data type for the status codes that you specify that, when returned by a primary origin, trigger CloudFront to failover to a second origin.

See: newStatusCodes smart constructor.

Constructors

StatusCodes'

Fields

quantity :: Int

The number of status codes.

items :: NonEmpty Int

The items (status codes) for an origin group.

newStatusCodes

Arguments

:: Int
$sel:quantity:StatusCodes'

-> NonEmpty Int
$sel:items:StatusCodes'

-> StatusCodes

Create a value of StatusCodes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quantity:StatusCodes', statusCodes_quantity - The number of status codes.

$sel:items:StatusCodes', statusCodes_items - The items (status codes) for an origin group.

statusCodes_quantity :: Lens' StatusCodes Int

The number of status codes.

statusCodes_items :: Lens' StatusCodes (NonEmpty Int)

The items (status codes) for an origin group.