Network.AWS.CloudFront.Types.StreamingDistributionConfig

Description

Documentation

data StreamingDistributionConfig

The RTMP distribution's configuration information.

See: newStreamingDistributionConfig smart constructor.

Constructors

StreamingDistributionConfig'

Fields

priceClass :: Maybe PriceClass

A complex type that contains information about price class for this streaming distribution.

logging :: Maybe StreamingLoggingConfig

A complex type that controls whether access logs are written for the streaming distribution.

aliases :: Maybe Aliases

A complex type that contains information about CNAMEs (alternate domain names), if any, for this streaming distribution.

callerReference :: Text

A unique value (for example, a date-time stamp) that ensures that the request can't be replayed.

If the value of CallerReference is new (regardless of the content of the StreamingDistributionConfig object), CloudFront creates a new distribution.

If CallerReference is a value that you already sent in a previous request to create a distribution, CloudFront returns a DistributionAlreadyExists error.

s3Origin :: S3Origin

A complex type that contains information about the Amazon S3 bucket from which you want CloudFront to get your media files for distribution.

comment :: Text

Any comments you want to include about the streaming distribution.

trustedSigners :: TrustedSigners

A complex type that specifies any accounts that you want to permit to create signed URLs for private content. If you want the distribution to use signed URLs, include this element; if you want the distribution to use public URLs, remove this element. For more information, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

enabled :: Bool

Whether the streaming distribution is enabled to accept user requests for content.

newStreamingDistributionConfig

Arguments

:: Text
$sel:callerReference:StreamingDistributionConfig'

-> S3Origin
$sel:s3Origin:StreamingDistributionConfig'

-> Text
$sel:comment:StreamingDistributionConfig'

-> TrustedSigners
$sel:trustedSigners:StreamingDistributionConfig'

-> Bool
$sel:enabled:StreamingDistributionConfig'

-> StreamingDistributionConfig

Create a value of StreamingDistributionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:priceClass:StreamingDistributionConfig', streamingDistributionConfig_priceClass - A complex type that contains information about price class for this streaming distribution.

$sel:logging:StreamingDistributionConfig', streamingDistributionConfig_logging - A complex type that controls whether access logs are written for the streaming distribution.

$sel:aliases:StreamingDistributionConfig', streamingDistributionConfig_aliases - A complex type that contains information about CNAMEs (alternate domain names), if any, for this streaming distribution.

$sel:callerReference:StreamingDistributionConfig', streamingDistributionConfig_callerReference - A unique value (for example, a date-time stamp) that ensures that the request can't be replayed.

If the value of CallerReference is new (regardless of the content of the StreamingDistributionConfig object), CloudFront creates a new distribution.

If CallerReference is a value that you already sent in a previous request to create a distribution, CloudFront returns a DistributionAlreadyExists error.

$sel:s3Origin:StreamingDistributionConfig', streamingDistributionConfig_s3Origin - A complex type that contains information about the Amazon S3 bucket from which you want CloudFront to get your media files for distribution.

$sel:comment:StreamingDistributionConfig', streamingDistributionConfig_comment - Any comments you want to include about the streaming distribution.

$sel:trustedSigners:StreamingDistributionConfig', streamingDistributionConfig_trustedSigners - A complex type that specifies any accounts that you want to permit to create signed URLs for private content. If you want the distribution to use signed URLs, include this element; if you want the distribution to use public URLs, remove this element. For more information, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

$sel:enabled:StreamingDistributionConfig', streamingDistributionConfig_enabled - Whether the streaming distribution is enabled to accept user requests for content.

streamingDistributionConfig_priceClass :: Lens' StreamingDistributionConfig (Maybe PriceClass)

A complex type that contains information about price class for this streaming distribution.

streamingDistributionConfig_logging :: Lens' StreamingDistributionConfig (Maybe StreamingLoggingConfig)

A complex type that controls whether access logs are written for the streaming distribution.

streamingDistributionConfig_aliases :: Lens' StreamingDistributionConfig (Maybe Aliases)

A complex type that contains information about CNAMEs (alternate domain names), if any, for this streaming distribution.

streamingDistributionConfig_callerReference :: Lens' StreamingDistributionConfig Text

A unique value (for example, a date-time stamp) that ensures that the request can't be replayed.

If the value of CallerReference is new (regardless of the content of the StreamingDistributionConfig object), CloudFront creates a new distribution.

If CallerReference is a value that you already sent in a previous request to create a distribution, CloudFront returns a DistributionAlreadyExists error.

streamingDistributionConfig_s3Origin :: Lens' StreamingDistributionConfig S3Origin

A complex type that contains information about the Amazon S3 bucket from which you want CloudFront to get your media files for distribution.

streamingDistributionConfig_comment :: Lens' StreamingDistributionConfig Text

Any comments you want to include about the streaming distribution.

streamingDistributionConfig_trustedSigners :: Lens' StreamingDistributionConfig TrustedSigners

A complex type that specifies any accounts that you want to permit to create signed URLs for private content. If you want the distribution to use signed URLs, include this element; if you want the distribution to use public URLs, remove this element. For more information, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

streamingDistributionConfig_enabled :: Lens' StreamingDistributionConfig Bool

Whether the streaming distribution is enabled to accept user requests for content.