Network.AWS.CloudFront.Types.StreamingDistributionList

Description

Documentation

data StreamingDistributionList

A streaming distribution list.

See: newStreamingDistributionList smart constructor.

Constructors

StreamingDistributionList'

Fields

items :: Maybe [StreamingDistributionSummary]

A complex type that contains one StreamingDistributionSummary element for each distribution that was created by the current account.

nextMarker :: Maybe Text

If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your RTMP distributions where they left off.

marker :: Text

The value you provided for the Marker request parameter.

maxItems :: Int

The value you provided for the MaxItems request parameter.

isTruncated :: Bool

A flag that indicates whether more streaming distributions remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more distributions in the list.

quantity :: Int

The number of streaming distributions that were created by the current account.

newStreamingDistributionList

Arguments

:: Text
$sel:marker:StreamingDistributionList'

-> Int
$sel:maxItems:StreamingDistributionList'

-> Bool
$sel:isTruncated:StreamingDistributionList'

-> Int
$sel:quantity:StreamingDistributionList'

-> StreamingDistributionList

Create a value of StreamingDistributionList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:StreamingDistributionList', streamingDistributionList_items - A complex type that contains one StreamingDistributionSummary element for each distribution that was created by the current account.

$sel:nextMarker:StreamingDistributionList', streamingDistributionList_nextMarker - If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your RTMP distributions where they left off.

$sel:marker:StreamingDistributionList', streamingDistributionList_marker - The value you provided for the Marker request parameter.

$sel:maxItems:StreamingDistributionList', streamingDistributionList_maxItems - The value you provided for the MaxItems request parameter.

$sel:isTruncated:StreamingDistributionList', streamingDistributionList_isTruncated - A flag that indicates whether more streaming distributions remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more distributions in the list.

$sel:quantity:StreamingDistributionList', streamingDistributionList_quantity - The number of streaming distributions that were created by the current account.

streamingDistributionList_items :: Lens' StreamingDistributionList (Maybe [StreamingDistributionSummary])

A complex type that contains one StreamingDistributionSummary element for each distribution that was created by the current account.

streamingDistributionList_nextMarker :: Lens' StreamingDistributionList (Maybe Text)

If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your RTMP distributions where they left off.

streamingDistributionList_marker :: Lens' StreamingDistributionList Text

The value you provided for the Marker request parameter.

streamingDistributionList_maxItems :: Lens' StreamingDistributionList Int

The value you provided for the MaxItems request parameter.

streamingDistributionList_isTruncated :: Lens' StreamingDistributionList Bool

A flag that indicates whether more streaming distributions remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more distributions in the list.

streamingDistributionList_quantity :: Lens' StreamingDistributionList Int

The number of streaming distributions that were created by the current account.