Network.AWS.CloudFront.Types.StreamingDistributionSummary

Description

Documentation

data StreamingDistributionSummary

A summary of the information for a CloudFront streaming distribution.

See: newStreamingDistributionSummary smart constructor.

Constructors

StreamingDistributionSummary'

Fields

id :: Text

The identifier for the distribution, for example, EDFDVBD632BHDS5.

arn :: Text

The ARN (Amazon Resource Name) for the streaming distribution. For example: arn:aws:cloudfront::123456789012:streaming-distribution/EDFDVBD632BHDS5, where 123456789012 is your account ID.

status :: Text

Indicates the current status of the distribution. When the status is Deployed, the distribution's information is fully propagated throughout the Amazon CloudFront system.

lastModifiedTime :: ISO8601

The date and time the distribution was last modified.

domainName :: Text

The domain name corresponding to the distribution, for example, d111111abcdef8.cloudfront.net.

s3Origin :: S3Origin

A complex type that contains information about the Amazon S3 bucket from which you want CloudFront to get your media files for distribution.

aliases :: Aliases

A complex type that contains information about CNAMEs (alternate domain names), if any, for this streaming distribution.

trustedSigners :: TrustedSigners

A complex type that specifies the accounts, if any, that you want to allow to create signed URLs for private content. If you want to require signed URLs in requests for objects in the target origin that match the PathPattern for this cache behavior, specify true for Enabled, and specify the applicable values for Quantity and Items.If you don't want to require signed URLs in requests for objects that match PathPattern, specify false for Enabled and 0 for Quantity. Omit Items. To add, change, or remove one or more trusted signers, change Enabled to true (if it's currently false), change Quantity as applicable, and specify all of the trusted signers that you want to include in the updated distribution.

For more information, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

comment :: Text

The comment originally specified when this distribution was created.

priceClass :: PriceClass

A complex type that contains information about price class for this streaming distribution.

enabled :: Bool

Whether the distribution is enabled to accept end user requests for content.

newStreamingDistributionSummary

Arguments

:: Text
$sel:id:StreamingDistributionSummary'

-> Text
$sel:arn:StreamingDistributionSummary'

-> Text
$sel:status:StreamingDistributionSummary'

-> UTCTime
$sel:lastModifiedTime:StreamingDistributionSummary'

-> Text
$sel:domainName:StreamingDistributionSummary'

-> S3Origin
$sel:s3Origin:StreamingDistributionSummary'

-> Aliases
$sel:aliases:StreamingDistributionSummary'

-> TrustedSigners
$sel:trustedSigners:StreamingDistributionSummary'

-> Text
$sel:comment:StreamingDistributionSummary'

-> PriceClass
$sel:priceClass:StreamingDistributionSummary'

-> Bool
$sel:enabled:StreamingDistributionSummary'

-> StreamingDistributionSummary

Create a value of StreamingDistributionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:StreamingDistributionSummary', streamingDistributionSummary_id - The identifier for the distribution, for example, EDFDVBD632BHDS5.

$sel:arn:StreamingDistributionSummary', streamingDistributionSummary_arn - The ARN (Amazon Resource Name) for the streaming distribution. For example: arn:aws:cloudfront::123456789012:streaming-distribution/EDFDVBD632BHDS5, where 123456789012 is your account ID.

$sel:status:StreamingDistributionSummary', streamingDistributionSummary_status - Indicates the current status of the distribution. When the status is Deployed, the distribution's information is fully propagated throughout the Amazon CloudFront system.

$sel:lastModifiedTime:StreamingDistributionSummary', streamingDistributionSummary_lastModifiedTime - The date and time the distribution was last modified.

$sel:domainName:StreamingDistributionSummary', streamingDistributionSummary_domainName - The domain name corresponding to the distribution, for example, d111111abcdef8.cloudfront.net.

$sel:s3Origin:StreamingDistributionSummary', streamingDistributionSummary_s3Origin - A complex type that contains information about the Amazon S3 bucket from which you want CloudFront to get your media files for distribution.

$sel:aliases:StreamingDistributionSummary', streamingDistributionSummary_aliases - A complex type that contains information about CNAMEs (alternate domain names), if any, for this streaming distribution.

$sel:trustedSigners:StreamingDistributionSummary', streamingDistributionSummary_trustedSigners - A complex type that specifies the accounts, if any, that you want to allow to create signed URLs for private content. If you want to require signed URLs in requests for objects in the target origin that match the PathPattern for this cache behavior, specify true for Enabled, and specify the applicable values for Quantity and Items.If you don't want to require signed URLs in requests for objects that match PathPattern, specify false for Enabled and 0 for Quantity. Omit Items. To add, change, or remove one or more trusted signers, change Enabled to true (if it's currently false), change Quantity as applicable, and specify all of the trusted signers that you want to include in the updated distribution.

For more information, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

$sel:comment:StreamingDistributionSummary', streamingDistributionSummary_comment - The comment originally specified when this distribution was created.

$sel:priceClass:StreamingDistributionSummary', streamingDistributionSummary_priceClass - A complex type that contains information about price class for this streaming distribution.

$sel:enabled:StreamingDistributionSummary', streamingDistributionSummary_enabled - Whether the distribution is enabled to accept end user requests for content.

streamingDistributionSummary_id :: Lens' StreamingDistributionSummary Text

The identifier for the distribution, for example, EDFDVBD632BHDS5.

streamingDistributionSummary_arn :: Lens' StreamingDistributionSummary Text

The ARN (Amazon Resource Name) for the streaming distribution. For example: arn:aws:cloudfront::123456789012:streaming-distribution/EDFDVBD632BHDS5, where 123456789012 is your account ID.

streamingDistributionSummary_status :: Lens' StreamingDistributionSummary Text

Indicates the current status of the distribution. When the status is Deployed, the distribution's information is fully propagated throughout the Amazon CloudFront system.

streamingDistributionSummary_lastModifiedTime :: Lens' StreamingDistributionSummary UTCTime

The date and time the distribution was last modified.

streamingDistributionSummary_domainName :: Lens' StreamingDistributionSummary Text

The domain name corresponding to the distribution, for example, d111111abcdef8.cloudfront.net.

streamingDistributionSummary_s3Origin :: Lens' StreamingDistributionSummary S3Origin

A complex type that contains information about the Amazon S3 bucket from which you want CloudFront to get your media files for distribution.

streamingDistributionSummary_aliases :: Lens' StreamingDistributionSummary Aliases

A complex type that contains information about CNAMEs (alternate domain names), if any, for this streaming distribution.

streamingDistributionSummary_trustedSigners :: Lens' StreamingDistributionSummary TrustedSigners

A complex type that specifies the accounts, if any, that you want to allow to create signed URLs for private content. If you want to require signed URLs in requests for objects in the target origin that match the PathPattern for this cache behavior, specify true for Enabled, and specify the applicable values for Quantity and Items.If you don't want to require signed URLs in requests for objects that match PathPattern, specify false for Enabled and 0 for Quantity. Omit Items. To add, change, or remove one or more trusted signers, change Enabled to true (if it's currently false), change Quantity as applicable, and specify all of the trusted signers that you want to include in the updated distribution.

For more information, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

streamingDistributionSummary_comment :: Lens' StreamingDistributionSummary Text

The comment originally specified when this distribution was created.

streamingDistributionSummary_priceClass :: Lens' StreamingDistributionSummary PriceClass

A complex type that contains information about price class for this streaming distribution.

streamingDistributionSummary_enabled :: Lens' StreamingDistributionSummary Bool

Whether the distribution is enabled to accept end user requests for content.