Network.AWS.CloudFront.Types.Tags

Description

Documentation

data Tags

A complex type that contains zero or more Tag elements.

See: newTags smart constructor.

Constructors

Tags'

Fields

items :: Maybe [Tag]

A complex type that contains Tag elements.

newTags :: Tags

Create a value of Tags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:Tags', tags_items - A complex type that contains Tag elements.

tags_items :: Lens' Tags (Maybe [Tag])

A complex type that contains Tag elements.