Network.AWS.CloudFront.Types.TestResult

Description

Documentation

data TestResult

Contains the result of testing a CloudFront function with TestFunction.

See: newTestResult smart constructor.

Constructors

TestResult'

Fields

computeUtilization :: Maybe Text

The amount of time that the function took to run as a percentage of the maximum allowed time. For example, a compute utilization of 35 means that the function completed in 35% of the maximum allowed time.

functionOutput :: Maybe Text

The event object returned by the function. For more information about the structure of the event object, see Event object structure in the Amazon CloudFront Developer Guide.

functionExecutionLogs :: Maybe [Text]

Contains the log lines that the function wrote (if any) when running the test.

functionErrorMessage :: Maybe Text

If the result of testing the function was an error, this field contains the error message.

functionSummary :: Maybe FunctionSummary

Contains configuration information and metadata about the CloudFront function that was tested.

newTestResult :: TestResult

Create a value of TestResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:computeUtilization:TestResult', testResult_computeUtilization - The amount of time that the function took to run as a percentage of the maximum allowed time. For example, a compute utilization of 35 means that the function completed in 35% of the maximum allowed time.

$sel:functionOutput:TestResult', testResult_functionOutput - The event object returned by the function. For more information about the structure of the event object, see Event object structure in the Amazon CloudFront Developer Guide.

$sel:functionExecutionLogs:TestResult', testResult_functionExecutionLogs - Contains the log lines that the function wrote (if any) when running the test.

$sel:functionErrorMessage:TestResult', testResult_functionErrorMessage - If the result of testing the function was an error, this field contains the error message.

$sel:functionSummary:TestResult', testResult_functionSummary - Contains configuration information and metadata about the CloudFront function that was tested.

testResult_computeUtilization :: Lens' TestResult (Maybe Text)

The amount of time that the function took to run as a percentage of the maximum allowed time. For example, a compute utilization of 35 means that the function completed in 35% of the maximum allowed time.

testResult_functionOutput :: Lens' TestResult (Maybe Text)

The event object returned by the function. For more information about the structure of the event object, see Event object structure in the Amazon CloudFront Developer Guide.

testResult_functionExecutionLogs :: Lens' TestResult (Maybe [Text])

Contains the log lines that the function wrote (if any) when running the test.

testResult_functionErrorMessage :: Lens' TestResult (Maybe Text)

If the result of testing the function was an error, this field contains the error message.

testResult_functionSummary :: Lens' TestResult (Maybe FunctionSummary)

Contains configuration information and metadata about the CloudFront function that was tested.