Network.AWS.CloudFront.UpdateCachePolicy

Description

Updates a cache policy configuration.

When you update a cache policy configuration, all the fields are updated with the values provided in the request. You cannot update some fields independent of others. To update a cache policy configuration:

Use GetCachePolicyConfig to get the current configuration.

Locally modify the fields in the cache policy configuration that you want to update.

Call UpdateCachePolicy by providing the entire cache policy configuration, including the fields that you modified and those that you didn’t.

Creating a Request

data UpdateCachePolicy

See: newUpdateCachePolicy smart constructor.

Constructors

UpdateCachePolicy'

Fields

ifMatch :: Maybe Text

The version of the cache policy that you are updating. The version is returned in the cache policy’s ETag field in the response to GetCachePolicyConfig.

cachePolicyConfig :: CachePolicyConfig

A cache policy configuration.

id :: Text

The unique identifier for the cache policy that you are updating. The identifier is returned in a cache behavior’s CachePolicyId field in the response to GetDistributionConfig.

newUpdateCachePolicy

Arguments

:: CachePolicyConfig
$sel:cachePolicyConfig:UpdateCachePolicy'

-> Text
$sel:id:UpdateCachePolicy'

-> UpdateCachePolicy

Create a value of UpdateCachePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifMatch:UpdateCachePolicy', updateCachePolicy_ifMatch - The version of the cache policy that you are updating. The version is returned in the cache policy’s ETag field in the response to GetCachePolicyConfig.

$sel:cachePolicyConfig:UpdateCachePolicy', updateCachePolicy_cachePolicyConfig - A cache policy configuration.

$sel:id:UpdateCachePolicy', updateCachePolicy_id - The unique identifier for the cache policy that you are updating. The identifier is returned in a cache behavior’s CachePolicyId field in the response to GetDistributionConfig.

Request Lenses

updateCachePolicy_ifMatch :: Lens' UpdateCachePolicy (Maybe Text)

The version of the cache policy that you are updating. The version is returned in the cache policy’s ETag field in the response to GetCachePolicyConfig.

updateCachePolicy_cachePolicyConfig :: Lens' UpdateCachePolicy CachePolicyConfig

A cache policy configuration.

updateCachePolicy_id :: Lens' UpdateCachePolicy Text

The unique identifier for the cache policy that you are updating. The identifier is returned in a cache behavior’s CachePolicyId field in the response to GetDistributionConfig.

Destructuring the Response

data UpdateCachePolicyResponse

See: newUpdateCachePolicyResponse smart constructor.

Constructors

UpdateCachePolicyResponse'

Fields

cachePolicy :: Maybe CachePolicy

A cache policy.

eTag :: Maybe Text

The current version of the cache policy.

httpStatus :: Int

The response's http status code.

newUpdateCachePolicyResponse

Arguments

:: Int
$sel:httpStatus:UpdateCachePolicyResponse'

-> UpdateCachePolicyResponse

Create a value of UpdateCachePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cachePolicy:UpdateCachePolicyResponse', updateCachePolicyResponse_cachePolicy - A cache policy.

$sel:eTag:UpdateCachePolicyResponse', updateCachePolicyResponse_eTag - The current version of the cache policy.

$sel:httpStatus:UpdateCachePolicyResponse', updateCachePolicyResponse_httpStatus - The response's http status code.

Response Lenses

updateCachePolicyResponse_cachePolicy :: Lens' UpdateCachePolicyResponse (Maybe CachePolicy)

A cache policy.

updateCachePolicyResponse_eTag :: Lens' UpdateCachePolicyResponse (Maybe Text)

The current version of the cache policy.

updateCachePolicyResponse_httpStatus :: Lens' UpdateCachePolicyResponse Int

The response's http status code.