Network.AWS.CloudFront.UpdateFieldLevelEncryptionConfig

Description

Update a field-level encryption configuration.

Creating a Request

data UpdateFieldLevelEncryptionConfig

See: newUpdateFieldLevelEncryptionConfig smart constructor.

Constructors

UpdateFieldLevelEncryptionConfig'

Fields

ifMatch :: Maybe Text

The value of the ETag header that you received when retrieving the configuration identity to update. For example: E2QWRUHAPOMQZL.

fieldLevelEncryptionConfig :: FieldLevelEncryptionConfig

Request to update a field-level encryption configuration.

id :: Text

The ID of the configuration you want to update.

newUpdateFieldLevelEncryptionConfig

Arguments

:: FieldLevelEncryptionConfig
$sel:fieldLevelEncryptionConfig:UpdateFieldLevelEncryptionConfig'

-> Text
$sel:id:UpdateFieldLevelEncryptionConfig'

-> UpdateFieldLevelEncryptionConfig

Create a value of UpdateFieldLevelEncryptionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifMatch:UpdateFieldLevelEncryptionConfig', updateFieldLevelEncryptionConfig_ifMatch - The value of the ETag header that you received when retrieving the configuration identity to update. For example: E2QWRUHAPOMQZL.

$sel:fieldLevelEncryptionConfig:UpdateFieldLevelEncryptionConfig', updateFieldLevelEncryptionConfig_fieldLevelEncryptionConfig - Request to update a field-level encryption configuration.

$sel:id:UpdateFieldLevelEncryptionConfig', updateFieldLevelEncryptionConfig_id - The ID of the configuration you want to update.

Request Lenses

updateFieldLevelEncryptionConfig_ifMatch :: Lens' UpdateFieldLevelEncryptionConfig (Maybe Text)

The value of the ETag header that you received when retrieving the configuration identity to update. For example: E2QWRUHAPOMQZL.

updateFieldLevelEncryptionConfig_fieldLevelEncryptionConfig :: Lens' UpdateFieldLevelEncryptionConfig FieldLevelEncryptionConfig

Request to update a field-level encryption configuration.

updateFieldLevelEncryptionConfig_id :: Lens' UpdateFieldLevelEncryptionConfig Text

The ID of the configuration you want to update.

Destructuring the Response

data UpdateFieldLevelEncryptionConfigResponse

See: newUpdateFieldLevelEncryptionConfigResponse smart constructor.

Constructors

UpdateFieldLevelEncryptionConfigResponse'

Fields

eTag :: Maybe Text

The value of the ETag header that you received when updating the configuration. For example: E2QWRUHAPOMQZL.

fieldLevelEncryption :: Maybe FieldLevelEncryption

Return the results of updating the configuration.

httpStatus :: Int

The response's http status code.

newUpdateFieldLevelEncryptionConfigResponse

Arguments

:: Int
$sel:httpStatus:UpdateFieldLevelEncryptionConfigResponse'

-> UpdateFieldLevelEncryptionConfigResponse

Create a value of UpdateFieldLevelEncryptionConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:UpdateFieldLevelEncryptionConfigResponse', updateFieldLevelEncryptionConfigResponse_eTag - The value of the ETag header that you received when updating the configuration. For example: E2QWRUHAPOMQZL.

$sel:fieldLevelEncryption:UpdateFieldLevelEncryptionConfigResponse', updateFieldLevelEncryptionConfigResponse_fieldLevelEncryption - Return the results of updating the configuration.

$sel:httpStatus:UpdateFieldLevelEncryptionConfigResponse', updateFieldLevelEncryptionConfigResponse_httpStatus - The response's http status code.

Response Lenses

updateFieldLevelEncryptionConfigResponse_eTag :: Lens' UpdateFieldLevelEncryptionConfigResponse (Maybe Text)

The value of the ETag header that you received when updating the configuration. For example: E2QWRUHAPOMQZL.

updateFieldLevelEncryptionConfigResponse_fieldLevelEncryption :: Lens' UpdateFieldLevelEncryptionConfigResponse (Maybe FieldLevelEncryption)

Return the results of updating the configuration.

updateFieldLevelEncryptionConfigResponse_httpStatus :: Lens' UpdateFieldLevelEncryptionConfigResponse Int

The response's http status code.