Network.AWS.CloudFront.UpdateFunction

Description

Updates a CloudFront function.

You can update a function’s code or the comment that describes the function. You cannot update a function’s name.

To update a function, you provide the function’s name and version (ETag value) along with the updated function code. To get the name and version, you can use ListFunctions and DescribeFunction.

Creating a Request

data UpdateFunction

See: newUpdateFunction smart constructor.

Constructors

UpdateFunction'

Fields

ifMatch :: Text

The current version (ETag value) of the function that you are updating, which you can get using DescribeFunction.

functionConfig :: FunctionConfig

Configuration information about the function.

functionCode :: Sensitive Base64

The function code. For more information about writing a CloudFront function, see Writing function code for CloudFront Functions in the Amazon CloudFront Developer Guide.

name :: Text

The name of the function that you are updating.

newUpdateFunction

Arguments

:: Text
$sel:ifMatch:UpdateFunction'

-> FunctionConfig
$sel:functionConfig:UpdateFunction'

-> ByteString
$sel:functionCode:UpdateFunction'

-> Text
$sel:name:UpdateFunction'

-> UpdateFunction

Create a value of UpdateFunction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifMatch:UpdateFunction', updateFunction_ifMatch - The current version (ETag value) of the function that you are updating, which you can get using DescribeFunction.

$sel:functionConfig:UpdateFunction', updateFunction_functionConfig - Configuration information about the function.

$sel:functionCode:UpdateFunction', updateFunction_functionCode - The function code. For more information about writing a CloudFront function, see Writing function code for CloudFront Functions in the Amazon CloudFront Developer Guide.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:name:UpdateFunction', updateFunction_name - The name of the function that you are updating.

Request Lenses

updateFunction_ifMatch :: Lens' UpdateFunction Text

The current version (ETag value) of the function that you are updating, which you can get using DescribeFunction.

updateFunction_functionConfig :: Lens' UpdateFunction FunctionConfig

Configuration information about the function.

updateFunction_functionCode :: Lens' UpdateFunction ByteString

The function code. For more information about writing a CloudFront function, see Writing function code for CloudFront Functions in the Amazon CloudFront Developer Guide.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

updateFunction_name :: Lens' UpdateFunction Text

The name of the function that you are updating.

Destructuring the Response

data UpdateFunctionResponse

See: newUpdateFunctionResponse smart constructor.

Constructors

UpdateFunctionResponse'

Fields

eTag :: Maybe Text

The version identifier for the current version of the CloudFront function.

functionSummary :: Maybe FunctionSummary

Contains configuration information and metadata about a CloudFront function.

httpStatus :: Int

The response's http status code.

newUpdateFunctionResponse

Arguments

:: Int
$sel:httpStatus:UpdateFunctionResponse'

-> UpdateFunctionResponse

Create a value of UpdateFunctionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:UpdateFunctionResponse', updateFunctionResponse_eTag - The version identifier for the current version of the CloudFront function.

$sel:functionSummary:UpdateFunctionResponse', updateFunctionResponse_functionSummary - Contains configuration information and metadata about a CloudFront function.

$sel:httpStatus:UpdateFunctionResponse', updateFunctionResponse_httpStatus - The response's http status code.

Response Lenses

updateFunctionResponse_eTag :: Lens' UpdateFunctionResponse (Maybe Text)

The version identifier for the current version of the CloudFront function.

updateFunctionResponse_functionSummary :: Lens' UpdateFunctionResponse (Maybe FunctionSummary)

Contains configuration information and metadata about a CloudFront function.

updateFunctionResponse_httpStatus :: Lens' UpdateFunctionResponse Int

The response's http status code.