Network.AWS.CloudFront.UpdateKeyGroup

Description

Updates a key group.

When you update a key group, all the fields are updated with the values provided in the request. You cannot update some fields independent of others. To update a key group:

Get the current key group with GetKeyGroup or GetKeyGroupConfig.

Locally modify the fields in the key group that you want to update. For example, add or remove public key IDs.

Call UpdateKeyGroup with the entire key group object, including the fields that you modified and those that you didn’t.

Creating a Request

data UpdateKeyGroup

See: newUpdateKeyGroup smart constructor.

Constructors

UpdateKeyGroup'

Fields

ifMatch :: Maybe Text

The version of the key group that you are updating. The version is the key group’s ETag value.

keyGroupConfig :: KeyGroupConfig

The key group configuration.

id :: Text

The identifier of the key group that you are updating.

newUpdateKeyGroup

Arguments

:: KeyGroupConfig
$sel:keyGroupConfig:UpdateKeyGroup'

-> Text
$sel:id:UpdateKeyGroup'

-> UpdateKeyGroup

Create a value of UpdateKeyGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ifMatch:UpdateKeyGroup', updateKeyGroup_ifMatch - The version of the key group that you are updating. The version is the key group’s ETag value.

$sel:keyGroupConfig:UpdateKeyGroup', updateKeyGroup_keyGroupConfig - The key group configuration.

$sel:id:UpdateKeyGroup', updateKeyGroup_id - The identifier of the key group that you are updating.

Request Lenses

updateKeyGroup_ifMatch :: Lens' UpdateKeyGroup (Maybe Text)

The version of the key group that you are updating. The version is the key group’s ETag value.

updateKeyGroup_keyGroupConfig :: Lens' UpdateKeyGroup KeyGroupConfig

The key group configuration.

updateKeyGroup_id :: Lens' UpdateKeyGroup Text

The identifier of the key group that you are updating.

Destructuring the Response

data UpdateKeyGroupResponse

See: newUpdateKeyGroupResponse smart constructor.

Constructors

UpdateKeyGroupResponse'

Fields

eTag :: Maybe Text

The identifier for this version of the key group.

keyGroup :: Maybe KeyGroup

The key group that was just updated.

httpStatus :: Int

The response's http status code.

newUpdateKeyGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateKeyGroupResponse'

-> UpdateKeyGroupResponse

Create a value of UpdateKeyGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:UpdateKeyGroupResponse', updateKeyGroupResponse_eTag - The identifier for this version of the key group.

$sel:keyGroup:UpdateKeyGroupResponse', updateKeyGroupResponse_keyGroup - The key group that was just updated.

$sel:httpStatus:UpdateKeyGroupResponse', updateKeyGroupResponse_httpStatus - The response's http status code.

Response Lenses

updateKeyGroupResponse_eTag :: Lens' UpdateKeyGroupResponse (Maybe Text)

The identifier for this version of the key group.

updateKeyGroupResponse_keyGroup :: Lens' UpdateKeyGroupResponse (Maybe KeyGroup)

The key group that was just updated.

updateKeyGroupResponse_httpStatus :: Lens' UpdateKeyGroupResponse Int

The response's http status code.