Network.AWS.CloudFront.UpdateRealtimeLogConfig

Description

Updates a real-time log configuration.

When you update a real-time log configuration, all the parameters are updated with the values provided in the request. You cannot update some parameters independent of others. To update a real-time log configuration:

Call GetRealtimeLogConfig to get the current real-time log configuration.

Locally modify the parameters in the real-time log configuration that you want to update.

Call this API (UpdateRealtimeLogConfig) by providing the entire real-time log configuration, including the parameters that you modified and those that you didn’t.

You cannot update a real-time log configuration’s Name or ARN.

Creating a Request

data UpdateRealtimeLogConfig

See: newUpdateRealtimeLogConfig smart constructor.

Constructors

UpdateRealtimeLogConfig'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) for this real-time log configuration.

samplingRate :: Maybe Integer

The sampling rate for this real-time log configuration. The sampling rate determines the percentage of viewer requests that are represented in the real-time log data. You must provide an integer between 1 and 100, inclusive.

name :: Maybe Text

The name for this real-time log configuration.

endPoints :: Maybe [EndPoint]

Contains information about the Amazon Kinesis data stream where you are sending real-time log data.

fields :: Maybe [Text]

A list of fields to include in each real-time log record.

For more information about fields, see Real-time log configuration fields in the Amazon CloudFront Developer Guide.

newUpdateRealtimeLogConfig :: UpdateRealtimeLogConfig

Create a value of UpdateRealtimeLogConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:UpdateRealtimeLogConfig', updateRealtimeLogConfig_arn - The Amazon Resource Name (ARN) for this real-time log configuration.

$sel:samplingRate:UpdateRealtimeLogConfig', updateRealtimeLogConfig_samplingRate - The sampling rate for this real-time log configuration. The sampling rate determines the percentage of viewer requests that are represented in the real-time log data. You must provide an integer between 1 and 100, inclusive.

$sel:name:UpdateRealtimeLogConfig', updateRealtimeLogConfig_name - The name for this real-time log configuration.

$sel:endPoints:UpdateRealtimeLogConfig', updateRealtimeLogConfig_endPoints - Contains information about the Amazon Kinesis data stream where you are sending real-time log data.

$sel:fields:UpdateRealtimeLogConfig', updateRealtimeLogConfig_fields - A list of fields to include in each real-time log record.

For more information about fields, see Real-time log configuration fields in the Amazon CloudFront Developer Guide.

Request Lenses

updateRealtimeLogConfig_arn :: Lens' UpdateRealtimeLogConfig (Maybe Text)

The Amazon Resource Name (ARN) for this real-time log configuration.

updateRealtimeLogConfig_samplingRate :: Lens' UpdateRealtimeLogConfig (Maybe Integer)

The sampling rate for this real-time log configuration. The sampling rate determines the percentage of viewer requests that are represented in the real-time log data. You must provide an integer between 1 and 100, inclusive.

updateRealtimeLogConfig_name :: Lens' UpdateRealtimeLogConfig (Maybe Text)

The name for this real-time log configuration.

updateRealtimeLogConfig_endPoints :: Lens' UpdateRealtimeLogConfig (Maybe [EndPoint])

Contains information about the Amazon Kinesis data stream where you are sending real-time log data.

updateRealtimeLogConfig_fields :: Lens' UpdateRealtimeLogConfig (Maybe [Text])

A list of fields to include in each real-time log record.

For more information about fields, see Real-time log configuration fields in the Amazon CloudFront Developer Guide.

Destructuring the Response

data UpdateRealtimeLogConfigResponse

See: newUpdateRealtimeLogConfigResponse smart constructor.

Constructors

UpdateRealtimeLogConfigResponse'

Fields

realtimeLogConfig :: Maybe RealtimeLogConfig

A real-time log configuration.

httpStatus :: Int

The response's http status code.

newUpdateRealtimeLogConfigResponse

Arguments

:: Int
$sel:httpStatus:UpdateRealtimeLogConfigResponse'

-> UpdateRealtimeLogConfigResponse

Create a value of UpdateRealtimeLogConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:realtimeLogConfig:UpdateRealtimeLogConfigResponse', updateRealtimeLogConfigResponse_realtimeLogConfig - A real-time log configuration.

$sel:httpStatus:UpdateRealtimeLogConfigResponse', updateRealtimeLogConfigResponse_httpStatus - The response's http status code.

Response Lenses

updateRealtimeLogConfigResponse_realtimeLogConfig :: Lens' UpdateRealtimeLogConfigResponse (Maybe RealtimeLogConfig)

A real-time log configuration.

updateRealtimeLogConfigResponse_httpStatus :: Lens' UpdateRealtimeLogConfigResponse Int

The response's http status code.