Network.AWS.CloudFront.V2014_05_31.ListDistributions

Description

List distributions.

Documentation

listDistributions :: ListDistributions

Minimum specification for a ListDistributions request.

data ListDistributions

Constructors

ListDistributions

Fields

_ldrMarker :: Maybe Text

Use this when paginating results to indicate where to begin in your list of distributions. The results include distributions in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last distribution on that page).

_ldrMaxItems :: Maybe Text

The maximum number of distributions you want in the response body.

ldrMaxItems :: Lens' ListDistributions (Maybe Text)

ldrMarker :: Lens' ListDistributions (Maybe Text)

data ListDistributionsResponse

Constructors

ListDistributionsResponse

Fields

_ldsDistributionList :: DistributionList

The DistributionList type.

ldsDistributionList :: Iso' ListDistributionsResponse DistributionList