Network.AWS.CloudHSM

Description

Derived from API version 2014-05-30 of the AWS service descriptions, licensed under Apache 2.0.

AWS CloudHSM Service

This is documentation for AWS CloudHSM Classic. For more information, see AWS CloudHSM Classic FAQs, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

For information about the current version of AWS CloudHSM, see AWS CloudHSM, the AWS CloudHSM User Guide, and the AWS CloudHSM API Reference.

Service Configuration

defaultService :: Service

API version 2014-05-30 of the Amazon CloudHSM SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CloudHSM.

CloudHsmInternalException

_CloudHsmInternalException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that an internal error occurred.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that one or more of the request parameters are not valid.

CloudHsmServiceException

_CloudHsmServiceException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that an exception occurred in the AWS CloudHSM service.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DeleteHapg

data DeleteHapg

Contains the inputs for the DeleteHapg action.

See: newDeleteHapg smart constructor.

Constructors

DeleteHapg' Text

newDeleteHapg

Arguments

:: Text
$sel:hapgArn:DeleteHapg'

-> DeleteHapg

Create a value of DeleteHapg with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hapgArn:DeleteHapg', deleteHapg_hapgArn - The ARN of the high-availability partition group to delete.

data DeleteHapgResponse

Contains the output of the DeleteHapg action.

See: newDeleteHapgResponse smart constructor.

Constructors

DeleteHapgResponse' Int Text

newDeleteHapgResponse

Arguments

:: Int
$sel:httpStatus:DeleteHapgResponse'

-> Text
$sel:status:DeleteHapgResponse'

-> DeleteHapgResponse

Create a value of DeleteHapgResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteHapgResponse', deleteHapgResponse_httpStatus - The response's http status code.

$sel:status:DeleteHapgResponse', deleteHapgResponse_status - The status of the action.

DeleteHsm

data DeleteHsm

Contains the inputs for the DeleteHsm operation.

See: newDeleteHsm smart constructor.

Constructors

DeleteHsm' Text

newDeleteHsm

Arguments

:: Text
$sel:hsmArn:DeleteHsm'

-> DeleteHsm

Create a value of DeleteHsm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsmArn:DeleteHsm', deleteHsm_hsmArn - The ARN of the HSM to delete.

data DeleteHsmResponse

Contains the output of the DeleteHsm operation.

See: newDeleteHsmResponse smart constructor.

Constructors

DeleteHsmResponse' Int Text

newDeleteHsmResponse

Arguments

:: Int
$sel:httpStatus:DeleteHsmResponse'

-> Text
$sel:status:DeleteHsmResponse'

-> DeleteHsmResponse

Create a value of DeleteHsmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteHsmResponse', deleteHsmResponse_httpStatus - The response's http status code.

$sel:status:DeleteHsmResponse', deleteHsmResponse_status - The status of the operation.

ModifyLunaClient

data ModifyLunaClient

See: newModifyLunaClient smart constructor.

Constructors

ModifyLunaClient' Text Text

newModifyLunaClient

Arguments

:: Text
$sel:clientArn:ModifyLunaClient'

-> Text
$sel:certificate:ModifyLunaClient'

-> ModifyLunaClient

Create a value of ModifyLunaClient with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientArn:ModifyLunaClient', modifyLunaClient_clientArn - The ARN of the client.

$sel:certificate:ModifyLunaClient', modifyLunaClient_certificate - The new certificate for the client.

data ModifyLunaClientResponse

See: newModifyLunaClientResponse smart constructor.

Constructors

ModifyLunaClientResponse' (Maybe Text) Int

newModifyLunaClientResponse

Arguments

:: Int
$sel:httpStatus:ModifyLunaClientResponse'

-> ModifyLunaClientResponse

Create a value of ModifyLunaClientResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientArn:ModifyLunaClient', modifyLunaClientResponse_clientArn - The ARN of the client.

$sel:httpStatus:ModifyLunaClientResponse', modifyLunaClientResponse_httpStatus - The response's http status code.

ListHsms (Paginated)

data ListHsms

See: newListHsms smart constructor.

Constructors

ListHsms' (Maybe Text)

newListHsms :: ListHsms

Create a value of ListHsms with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListHsms', listHsms_nextToken - The NextToken value from a previous call to ListHsms. Pass null if this is the first call.

data ListHsmsResponse

Contains the output of the ListHsms operation.

See: newListHsmsResponse smart constructor.

Constructors

ListHsmsResponse' (Maybe [Text]) (Maybe Text) Int

newListHsmsResponse

Arguments

:: Int
$sel:httpStatus:ListHsmsResponse'

-> ListHsmsResponse

Create a value of ListHsmsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsmList:ListHsmsResponse', listHsmsResponse_hsmList - The list of ARNs that identify the HSMs.

$sel:nextToken:ListHsms', listHsmsResponse_nextToken - If not null, more results are available. Pass this value to ListHsms to retrieve the next set of items.

$sel:httpStatus:ListHsmsResponse', listHsmsResponse_httpStatus - The response's http status code.

GetConfig

data GetConfig

See: newGetConfig smart constructor.

Constructors

GetConfig' Text ClientVersion [Text]

newGetConfig

Arguments

:: Text
$sel:clientArn:GetConfig'

-> ClientVersion
$sel:clientVersion:GetConfig'

-> GetConfig

Create a value of GetConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientArn:GetConfig', getConfig_clientArn - The ARN of the client.

$sel:clientVersion:GetConfig', getConfig_clientVersion - The client version.

$sel:hapgList:GetConfig', getConfig_hapgList - A list of ARNs that identify the high-availability partition groups that are associated with the client.

data GetConfigResponse

See: newGetConfigResponse smart constructor.

Constructors

GetConfigResponse' (Maybe Text) (Maybe Text) (Maybe Text) Int

newGetConfigResponse

Arguments

:: Int
$sel:httpStatus:GetConfigResponse'

-> GetConfigResponse

Create a value of GetConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configFile:GetConfigResponse', getConfigResponse_configFile - The chrystoki.conf configuration file.

$sel:configCred:GetConfigResponse', getConfigResponse_configCred - The certificate file containing the server.pem files of the HSMs.

$sel:configType:GetConfigResponse', getConfigResponse_configType - The type of credentials.

$sel:httpStatus:GetConfigResponse', getConfigResponse_httpStatus - The response's http status code.

DeleteLunaClient

data DeleteLunaClient

See: newDeleteLunaClient smart constructor.

Constructors

DeleteLunaClient' Text

newDeleteLunaClient

Arguments

:: Text
$sel:clientArn:DeleteLunaClient'

-> DeleteLunaClient

Create a value of DeleteLunaClient with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientArn:DeleteLunaClient', deleteLunaClient_clientArn - The ARN of the client to delete.

data DeleteLunaClientResponse

See: newDeleteLunaClientResponse smart constructor.

Constructors

DeleteLunaClientResponse' Int Text

newDeleteLunaClientResponse

Arguments

:: Int
$sel:httpStatus:DeleteLunaClientResponse'

-> Text
$sel:status:DeleteLunaClientResponse'

-> DeleteLunaClientResponse

Create a value of DeleteLunaClientResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteLunaClientResponse', deleteLunaClientResponse_httpStatus - The response's http status code.

$sel:status:DeleteLunaClientResponse', deleteLunaClientResponse_status - The status of the action.

ListAvailableZones

data ListAvailableZones

Contains the inputs for the ListAvailableZones action.

See: newListAvailableZones smart constructor.

Constructors

ListAvailableZones'

newListAvailableZones :: ListAvailableZones

Create a value of ListAvailableZones with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

data ListAvailableZonesResponse

See: newListAvailableZonesResponse smart constructor.

Constructors

ListAvailableZonesResponse' (Maybe [Text]) Int

newListAvailableZonesResponse

Arguments

:: Int
$sel:httpStatus:ListAvailableZonesResponse'

-> ListAvailableZonesResponse

Create a value of ListAvailableZonesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aZList:ListAvailableZonesResponse', listAvailableZonesResponse_aZList - The list of Availability Zones that have available AWS CloudHSM capacity.

$sel:httpStatus:ListAvailableZonesResponse', listAvailableZonesResponse_httpStatus - The response's http status code.

ModifyHapg

data ModifyHapg

See: newModifyHapg smart constructor.

Constructors

ModifyHapg' (Maybe [Text]) (Maybe Text) Text

newModifyHapg

Arguments

:: Text
$sel:hapgArn:ModifyHapg'

-> ModifyHapg

Create a value of ModifyHapg with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:partitionSerialList:ModifyHapg', modifyHapg_partitionSerialList - The list of partition serial numbers to make members of the high-availability partition group.

$sel:label:ModifyHapg', modifyHapg_label - The new label for the high-availability partition group.

$sel:hapgArn:ModifyHapg', modifyHapg_hapgArn - The ARN of the high-availability partition group to modify.

data ModifyHapgResponse

See: newModifyHapgResponse smart constructor.

Constructors

ModifyHapgResponse' (Maybe Text) Int

newModifyHapgResponse

Arguments

:: Int
$sel:httpStatus:ModifyHapgResponse'

-> ModifyHapgResponse

Create a value of ModifyHapgResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hapgArn:ModifyHapg', modifyHapgResponse_hapgArn - The ARN of the high-availability partition group.

$sel:httpStatus:ModifyHapgResponse', modifyHapgResponse_httpStatus - The response's http status code.

ListLunaClients (Paginated)

data ListLunaClients

See: newListLunaClients smart constructor.

Constructors

ListLunaClients' (Maybe Text)

newListLunaClients :: ListLunaClients

Create a value of ListLunaClients with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLunaClients', listLunaClients_nextToken - The NextToken value from a previous call to ListLunaClients. Pass null if this is the first call.

data ListLunaClientsResponse

See: newListLunaClientsResponse smart constructor.

Constructors

ListLunaClientsResponse' (Maybe Text) Int [Text]

newListLunaClientsResponse

Arguments

:: Int
$sel:httpStatus:ListLunaClientsResponse'

-> ListLunaClientsResponse

Create a value of ListLunaClientsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLunaClients', listLunaClientsResponse_nextToken - If not null, more results are available. Pass this to ListLunaClients to retrieve the next set of items.

$sel:httpStatus:ListLunaClientsResponse', listLunaClientsResponse_httpStatus - The response's http status code.

$sel:clientList:ListLunaClientsResponse', listLunaClientsResponse_clientList - The list of clients.

DescribeHsm

data DescribeHsm

Contains the inputs for the DescribeHsm operation.

See: newDescribeHsm smart constructor.

Constructors

DescribeHsm' (Maybe Text) (Maybe Text)

newDescribeHsm :: DescribeHsm

Create a value of DescribeHsm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsmSerialNumber:DescribeHsm', describeHsm_hsmSerialNumber - The serial number of the HSM. Either the HsmArn or the HsmSerialNumber parameter must be specified.

$sel:hsmArn:DescribeHsm', describeHsm_hsmArn - The ARN of the HSM. Either the HsmArn or the SerialNumber parameter must be specified.

data DescribeHsmResponse

Contains the output of the DescribeHsm operation.

See: newDescribeHsmResponse smart constructor.

Constructors

DescribeHsmResponse' (Maybe Text) (Maybe HsmStatus) (Maybe Text) (Maybe [Text]) (Maybe Text) (Maybe SubscriptionType) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Int

newDescribeHsmResponse

Arguments

:: Int
$sel:httpStatus:DescribeHsmResponse'

-> DescribeHsmResponse

Create a value of DescribeHsmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionStartDate:DescribeHsmResponse', describeHsmResponse_subscriptionStartDate - The subscription start date.

$sel:status:DescribeHsmResponse', describeHsmResponse_status - The status of the HSM.

$sel:iamRoleArn:DescribeHsmResponse', describeHsmResponse_iamRoleArn - The ARN of the IAM role assigned to the HSM.

$sel:partitions:DescribeHsmResponse', describeHsmResponse_partitions - The list of partitions on the HSM.

$sel:statusDetails:DescribeHsmResponse', describeHsmResponse_statusDetails - Contains additional information about the status of the HSM.

$sel:subscriptionType:DescribeHsmResponse', describeHsmResponse_subscriptionType - Undocumented member.

$sel:eniIp:DescribeHsmResponse', describeHsmResponse_eniIp - The IP address assigned to the HSM's ENI.

$sel:serverCertLastUpdated:DescribeHsmResponse', describeHsmResponse_serverCertLastUpdated - The date and time that the server certificate was last updated.

$sel:eniId:DescribeHsmResponse', describeHsmResponse_eniId - The identifier of the elastic network interface (ENI) attached to the HSM.

$sel:availabilityZone:DescribeHsmResponse', describeHsmResponse_availabilityZone - The Availability Zone that the HSM is in.

$sel:subnetId:DescribeHsmResponse', describeHsmResponse_subnetId - The identifier of the subnet that the HSM is in.

$sel:hsmType:DescribeHsmResponse', describeHsmResponse_hsmType - The HSM model type.

$sel:sshPublicKey:DescribeHsmResponse', describeHsmResponse_sshPublicKey - The public SSH key.

$sel:hsmArn:DescribeHsm', describeHsmResponse_hsmArn - The ARN of the HSM.

$sel:serverCertUri:DescribeHsmResponse', describeHsmResponse_serverCertUri - The URI of the certificate server.

$sel:serialNumber:DescribeHsmResponse', describeHsmResponse_serialNumber - The serial number of the HSM.

$sel:vendorName:DescribeHsmResponse', describeHsmResponse_vendorName - The name of the HSM vendor.

$sel:subscriptionEndDate:DescribeHsmResponse', describeHsmResponse_subscriptionEndDate - The subscription end date.

$sel:vpcId:DescribeHsmResponse', describeHsmResponse_vpcId - The identifier of the VPC that the HSM is in.

$sel:sshKeyLastUpdated:DescribeHsmResponse', describeHsmResponse_sshKeyLastUpdated - The date and time that the SSH key was last updated.

$sel:softwareVersion:DescribeHsmResponse', describeHsmResponse_softwareVersion - The HSM software version.

$sel:httpStatus:DescribeHsmResponse', describeHsmResponse_httpStatus - The response's http status code.

RemoveTagsFromResource

data RemoveTagsFromResource

See: newRemoveTagsFromResource smart constructor.

Constructors

RemoveTagsFromResource' Text [Text]

newRemoveTagsFromResource

Arguments

:: Text
$sel:resourceArn:RemoveTagsFromResource'

-> RemoveTagsFromResource

Create a value of RemoveTagsFromResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:RemoveTagsFromResource', removeTagsFromResource_resourceArn - The Amazon Resource Name (ARN) of the AWS CloudHSM resource.

$sel:tagKeyList:RemoveTagsFromResource', removeTagsFromResource_tagKeyList - The tag key or keys to remove.

Specify only the tag key to remove (not the value). To overwrite the value for an existing tag, use AddTagsToResource.

data RemoveTagsFromResourceResponse

See: newRemoveTagsFromResourceResponse smart constructor.

Constructors

RemoveTagsFromResourceResponse' Int Text

newRemoveTagsFromResourceResponse

Arguments

:: Int
$sel:httpStatus:RemoveTagsFromResourceResponse'

-> Text
$sel:status:RemoveTagsFromResourceResponse'

-> RemoveTagsFromResourceResponse

Create a value of RemoveTagsFromResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RemoveTagsFromResourceResponse', removeTagsFromResourceResponse_httpStatus - The response's http status code.

$sel:status:RemoveTagsFromResourceResponse', removeTagsFromResourceResponse_status - The status of the operation.

CreateLunaClient

data CreateLunaClient

Contains the inputs for the CreateLunaClient action.

See: newCreateLunaClient smart constructor.

Constructors

CreateLunaClient' (Maybe Text) Text

newCreateLunaClient

Arguments

:: Text
$sel:certificate:CreateLunaClient'

-> CreateLunaClient

Create a value of CreateLunaClient with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:label:CreateLunaClient', createLunaClient_label - The label for the client.

$sel:certificate:CreateLunaClient', createLunaClient_certificate - The contents of a Base64-Encoded X.509 v3 certificate to be installed on the HSMs used by this client.

data CreateLunaClientResponse

Contains the output of the CreateLunaClient action.

See: newCreateLunaClientResponse smart constructor.

Constructors

CreateLunaClientResponse' (Maybe Text) Int

newCreateLunaClientResponse

Arguments

:: Int
$sel:httpStatus:CreateLunaClientResponse'

-> CreateLunaClientResponse

Create a value of CreateLunaClientResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientArn:CreateLunaClientResponse', createLunaClientResponse_clientArn - The ARN of the client.

$sel:httpStatus:CreateLunaClientResponse', createLunaClientResponse_httpStatus - The response's http status code.

DescribeHapg

data DescribeHapg

Contains the inputs for the DescribeHapg action.

See: newDescribeHapg smart constructor.

Constructors

DescribeHapg' Text

newDescribeHapg

Arguments

:: Text
$sel:hapgArn:DescribeHapg'

-> DescribeHapg

Create a value of DescribeHapg with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hapgArn:DescribeHapg', describeHapg_hapgArn - The ARN of the high-availability partition group to describe.

data DescribeHapgResponse

Contains the output of the DescribeHapg action.

See: newDescribeHapgResponse smart constructor.

Constructors

DescribeHapgResponse' (Maybe [Text]) (Maybe Text) (Maybe [Text]) (Maybe Text) (Maybe CloudHsmObjectState) (Maybe Text) (Maybe Text) (Maybe [Text]) (Maybe [Text]) Int

newDescribeHapgResponse

Arguments

:: Int
$sel:httpStatus:DescribeHapgResponse'

-> DescribeHapgResponse

Create a value of DescribeHapgResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsmsPendingDeletion:DescribeHapgResponse', describeHapgResponse_hsmsPendingDeletion -

$sel:hapgArn:DescribeHapg', describeHapgResponse_hapgArn - The ARN of the high-availability partition group.

$sel:partitionSerialList:DescribeHapgResponse', describeHapgResponse_partitionSerialList - The list of partition serial numbers that belong to the high-availability partition group.

$sel:lastModifiedTimestamp:DescribeHapgResponse', describeHapgResponse_lastModifiedTimestamp - The date and time the high-availability partition group was last modified.

$sel:state:DescribeHapgResponse', describeHapgResponse_state - The state of the high-availability partition group.

$sel:hapgSerial:DescribeHapgResponse', describeHapgResponse_hapgSerial - The serial number of the high-availability partition group.

$sel:label:DescribeHapgResponse', describeHapgResponse_label - The label for the high-availability partition group.

$sel:hsmsPendingRegistration:DescribeHapgResponse', describeHapgResponse_hsmsPendingRegistration -

$sel:hsmsLastActionFailed:DescribeHapgResponse', describeHapgResponse_hsmsLastActionFailed -

$sel:httpStatus:DescribeHapgResponse', describeHapgResponse_httpStatus - The response's http status code.

CreateHapg

data CreateHapg

Contains the inputs for the CreateHapgRequest action.

See: newCreateHapg smart constructor.

Constructors

CreateHapg' Text

newCreateHapg

Arguments

:: Text
$sel:label:CreateHapg'

-> CreateHapg

Create a value of CreateHapg with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:label:CreateHapg', createHapg_label - The label of the new high-availability partition group.

data CreateHapgResponse

Contains the output of the CreateHAPartitionGroup action.

See: newCreateHapgResponse smart constructor.

Constructors

CreateHapgResponse' (Maybe Text) Int

newCreateHapgResponse

Arguments

:: Int
$sel:httpStatus:CreateHapgResponse'

-> CreateHapgResponse

Create a value of CreateHapgResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hapgArn:CreateHapgResponse', createHapgResponse_hapgArn - The ARN of the high-availability partition group.

$sel:httpStatus:CreateHapgResponse', createHapgResponse_httpStatus - The response's http status code.

DescribeLunaClient

data DescribeLunaClient

See: newDescribeLunaClient smart constructor.

Constructors

DescribeLunaClient' (Maybe Text) (Maybe Text)

newDescribeLunaClient :: DescribeLunaClient

Create a value of DescribeLunaClient with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientArn:DescribeLunaClient', describeLunaClient_clientArn - The ARN of the client.

$sel:certificateFingerprint:DescribeLunaClient', describeLunaClient_certificateFingerprint - The certificate fingerprint.

data DescribeLunaClientResponse

See: newDescribeLunaClientResponse smart constructor.

Constructors

DescribeLunaClientResponse' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Int

newDescribeLunaClientResponse

Arguments

:: Int
$sel:httpStatus:DescribeLunaClientResponse'

-> DescribeLunaClientResponse

Create a value of DescribeLunaClientResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTimestamp:DescribeLunaClientResponse', describeLunaClientResponse_lastModifiedTimestamp - The date and time the client was last modified.

$sel:clientArn:DescribeLunaClient', describeLunaClientResponse_clientArn - The ARN of the client.

$sel:label:DescribeLunaClientResponse', describeLunaClientResponse_label - The label of the client.

$sel:certificate:DescribeLunaClientResponse', describeLunaClientResponse_certificate - The certificate installed on the HSMs used by this client.

$sel:certificateFingerprint:DescribeLunaClient', describeLunaClientResponse_certificateFingerprint - The certificate fingerprint.

$sel:httpStatus:DescribeLunaClientResponse', describeLunaClientResponse_httpStatus - The response's http status code.

ListHapgs (Paginated)

data ListHapgs

See: newListHapgs smart constructor.

Constructors

ListHapgs' (Maybe Text)

newListHapgs :: ListHapgs

Create a value of ListHapgs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListHapgs', listHapgs_nextToken - The NextToken value from a previous call to ListHapgs. Pass null if this is the first call.

data ListHapgsResponse

See: newListHapgsResponse smart constructor.

Constructors

ListHapgsResponse' (Maybe Text) Int [Text]

newListHapgsResponse

Arguments

:: Int
$sel:httpStatus:ListHapgsResponse'

-> ListHapgsResponse

Create a value of ListHapgsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListHapgs', listHapgsResponse_nextToken - If not null, more results are available. Pass this value to ListHapgs to retrieve the next set of items.

$sel:httpStatus:ListHapgsResponse', listHapgsResponse_httpStatus - The response's http status code.

$sel:hapgList:ListHapgsResponse', listHapgsResponse_hapgList - The list of high-availability partition groups.

AddTagsToResource

data AddTagsToResource

See: newAddTagsToResource smart constructor.

Constructors

AddTagsToResource' Text [Tag]

newAddTagsToResource

Arguments

:: Text
$sel:resourceArn:AddTagsToResource'

-> AddTagsToResource

Create a value of AddTagsToResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:AddTagsToResource', addTagsToResource_resourceArn - The Amazon Resource Name (ARN) of the AWS CloudHSM resource to tag.

$sel:tagList:AddTagsToResource', addTagsToResource_tagList - One or more tags.

data AddTagsToResourceResponse

See: newAddTagsToResourceResponse smart constructor.

Constructors

AddTagsToResourceResponse' Int Text

newAddTagsToResourceResponse

Arguments

:: Int
$sel:httpStatus:AddTagsToResourceResponse'

-> Text
$sel:status:AddTagsToResourceResponse'

-> AddTagsToResourceResponse

Create a value of AddTagsToResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AddTagsToResourceResponse', addTagsToResourceResponse_httpStatus - The response's http status code.

$sel:status:AddTagsToResourceResponse', addTagsToResourceResponse_status - The status of the operation.

ModifyHsm

data ModifyHsm

Contains the inputs for the ModifyHsm operation.

See: newModifyHsm smart constructor.

Constructors

ModifyHsm' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Text

newModifyHsm

Arguments

:: Text
$sel:hsmArn:ModifyHsm'

-> ModifyHsm

Create a value of ModifyHsm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamRoleArn:ModifyHsm', modifyHsm_iamRoleArn - The new IAM role ARN.

$sel:eniIp:ModifyHsm', modifyHsm_eniIp - The new IP address for the elastic network interface (ENI) attached to the HSM.

If the HSM is moved to a different subnet, and an IP address is not specified, an IP address will be randomly chosen from the CIDR range of the new subnet.

$sel:syslogIp:ModifyHsm', modifyHsm_syslogIp - The new IP address for the syslog monitoring server. The AWS CloudHSM service only supports one syslog monitoring server.

$sel:subnetId:ModifyHsm', modifyHsm_subnetId - The new identifier of the subnet that the HSM is in. The new subnet must be in the same Availability Zone as the current subnet.

$sel:externalId:ModifyHsm', modifyHsm_externalId - The new external ID.

$sel:hsmArn:ModifyHsm', modifyHsm_hsmArn - The ARN of the HSM to modify.

data ModifyHsmResponse

Contains the output of the ModifyHsm operation.

See: newModifyHsmResponse smart constructor.

Constructors

ModifyHsmResponse' (Maybe Text) Int

newModifyHsmResponse

Arguments

:: Int
$sel:httpStatus:ModifyHsmResponse'

-> ModifyHsmResponse

Create a value of ModifyHsmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsmArn:ModifyHsm', modifyHsmResponse_hsmArn - The ARN of the HSM.

$sel:httpStatus:ModifyHsmResponse', modifyHsmResponse_httpStatus - The response's http status code.

CreateHsm

data CreateHsm

Contains the inputs for the CreateHsm operation.

See: newCreateHsm smart constructor.

Constructors

CreateHsm' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Text Text Text SubscriptionType

newCreateHsm

Arguments

:: Text
$sel:subnetId:CreateHsm'

-> Text
$sel:sshKey:CreateHsm'

-> Text
$sel:iamRoleArn:CreateHsm'

-> SubscriptionType
$sel:subscriptionType:CreateHsm'

-> CreateHsm

Create a value of CreateHsm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eniIp:CreateHsm', createHsm_eniIp - The IP address to assign to the HSM's ENI.

If an IP address is not specified, an IP address will be randomly chosen from the CIDR range of the subnet.

$sel:syslogIp:CreateHsm', createHsm_syslogIp - The IP address for the syslog monitoring server. The AWS CloudHSM service only supports one syslog monitoring server.

$sel:externalId:CreateHsm', createHsm_externalId - The external ID from IamRoleArn, if present.

$sel:clientToken:CreateHsm', createHsm_clientToken - A user-defined token to ensure idempotence. Subsequent calls to this operation with the same token will be ignored.

$sel:subnetId:CreateHsm', createHsm_subnetId - The identifier of the subnet in your VPC in which to place the HSM.

$sel:sshKey:CreateHsm', createHsm_sshKey - The SSH public key to install on the HSM.

$sel:iamRoleArn:CreateHsm', createHsm_iamRoleArn - The ARN of an IAM role to enable the AWS CloudHSM service to allocate an ENI on your behalf.

$sel:subscriptionType:CreateHsm', createHsm_subscriptionType - Undocumented member.

data CreateHsmResponse

Contains the output of the CreateHsm operation.

See: newCreateHsmResponse smart constructor.

Constructors

CreateHsmResponse' (Maybe Text) Int

newCreateHsmResponse

Arguments

:: Int
$sel:httpStatus:CreateHsmResponse'

-> CreateHsmResponse

Create a value of CreateHsmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsmArn:CreateHsmResponse', createHsmResponse_hsmArn - The ARN of the HSM.

$sel:httpStatus:CreateHsmResponse', createHsmResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The Amazon Resource Name (ARN) of the AWS CloudHSM resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' Int [Tag]

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

$sel:tagList:ListTagsForResourceResponse', listTagsForResourceResponse_tagList - One or more tags.

Types

ClientVersion

newtype ClientVersion

Constructors

ClientVersion'

Fields

fromClientVersion :: Text

Bundled Patterns

pattern ClientVersion_ :: ClientVersion

CloudHsmObjectState

newtype CloudHsmObjectState

Constructors

CloudHsmObjectState'

Fields

fromCloudHsmObjectState :: Text

Bundled Patterns

pattern CloudHsmObjectState_DEGRADED :: CloudHsmObjectState

pattern CloudHsmObjectState_READY :: CloudHsmObjectState

pattern CloudHsmObjectState_UPDATING :: CloudHsmObjectState

HsmStatus

newtype HsmStatus

Constructors

HsmStatus'

Fields

fromHsmStatus :: Text

Bundled Patterns

pattern HsmStatus_DEGRADED :: HsmStatus

pattern HsmStatus_PENDING :: HsmStatus

pattern HsmStatus_RUNNING :: HsmStatus

pattern HsmStatus_SUSPENDED :: HsmStatus

pattern HsmStatus_TERMINATED :: HsmStatus

pattern HsmStatus_TERMINATING :: HsmStatus

pattern HsmStatus_UPDATING :: HsmStatus

SubscriptionType

newtype SubscriptionType

Specifies the type of subscription for the HSM.

PRODUCTION - The HSM is being used in a production environment.

TRIAL - The HSM is being used in a product trial.

Constructors

SubscriptionType'

Fields

fromSubscriptionType :: Text

Bundled Patterns

pattern SubscriptionType_PRODUCTION :: SubscriptionType

Tag

data Tag

A key-value pair that identifies or specifies metadata about an AWS CloudHSM resource.

See: newTag smart constructor.

Constructors

Tag' Text Text

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key of the tag.

$sel:value:Tag', tag_value - The value of the tag.