Network.AWS.CloudHSM.DescribeHapg

Description

This is documentation for AWS CloudHSM Classic. For more information, see AWS CloudHSM Classic FAQs, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

For information about the current version of AWS CloudHSM, see AWS CloudHSM, the AWS CloudHSM User Guide, and the AWS CloudHSM API Reference.

Retrieves information about a high-availability partition group.

Creating a Request

data DescribeHapg

Contains the inputs for the DescribeHapg action.

See: newDescribeHapg smart constructor.

Constructors

DescribeHapg'

Fields

hapgArn :: Text

The ARN of the high-availability partition group to describe.

newDescribeHapg

Arguments

:: Text
$sel:hapgArn:DescribeHapg'

-> DescribeHapg

Create a value of DescribeHapg with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hapgArn:DescribeHapg', describeHapg_hapgArn - The ARN of the high-availability partition group to describe.

Request Lenses

describeHapg_hapgArn :: Lens' DescribeHapg Text

The ARN of the high-availability partition group to describe.

Destructuring the Response

data DescribeHapgResponse

Contains the output of the DescribeHapg action.

See: newDescribeHapgResponse smart constructor.

Constructors

DescribeHapgResponse'

Fields

hsmsPendingDeletion :: Maybe [Text]

hapgArn :: Maybe Text

The ARN of the high-availability partition group.

partitionSerialList :: Maybe [Text]

The list of partition serial numbers that belong to the high-availability partition group.

lastModifiedTimestamp :: Maybe Text

The date and time the high-availability partition group was last modified.

state :: Maybe CloudHsmObjectState

The state of the high-availability partition group.

hapgSerial :: Maybe Text

The serial number of the high-availability partition group.

label :: Maybe Text

The label for the high-availability partition group.

hsmsPendingRegistration :: Maybe [Text]

hsmsLastActionFailed :: Maybe [Text]

httpStatus :: Int

The response's http status code.

newDescribeHapgResponse

Arguments

:: Int
$sel:httpStatus:DescribeHapgResponse'

-> DescribeHapgResponse

Create a value of DescribeHapgResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsmsPendingDeletion:DescribeHapgResponse', describeHapgResponse_hsmsPendingDeletion -

$sel:hapgArn:DescribeHapg', describeHapgResponse_hapgArn - The ARN of the high-availability partition group.

$sel:partitionSerialList:DescribeHapgResponse', describeHapgResponse_partitionSerialList - The list of partition serial numbers that belong to the high-availability partition group.

$sel:lastModifiedTimestamp:DescribeHapgResponse', describeHapgResponse_lastModifiedTimestamp - The date and time the high-availability partition group was last modified.

$sel:state:DescribeHapgResponse', describeHapgResponse_state - The state of the high-availability partition group.

$sel:hapgSerial:DescribeHapgResponse', describeHapgResponse_hapgSerial - The serial number of the high-availability partition group.

$sel:label:DescribeHapgResponse', describeHapgResponse_label - The label for the high-availability partition group.

$sel:hsmsPendingRegistration:DescribeHapgResponse', describeHapgResponse_hsmsPendingRegistration -

$sel:hsmsLastActionFailed:DescribeHapgResponse', describeHapgResponse_hsmsLastActionFailed -

$sel:httpStatus:DescribeHapgResponse', describeHapgResponse_httpStatus - The response's http status code.

Response Lenses

describeHapgResponse_hsmsPendingDeletion :: Lens' DescribeHapgResponse (Maybe [Text])

describeHapgResponse_hapgArn :: Lens' DescribeHapgResponse (Maybe Text)

The ARN of the high-availability partition group.

describeHapgResponse_partitionSerialList :: Lens' DescribeHapgResponse (Maybe [Text])

The list of partition serial numbers that belong to the high-availability partition group.

describeHapgResponse_lastModifiedTimestamp :: Lens' DescribeHapgResponse (Maybe Text)

The date and time the high-availability partition group was last modified.

describeHapgResponse_state :: Lens' DescribeHapgResponse (Maybe CloudHsmObjectState)

The state of the high-availability partition group.

describeHapgResponse_hapgSerial :: Lens' DescribeHapgResponse (Maybe Text)

The serial number of the high-availability partition group.

describeHapgResponse_label :: Lens' DescribeHapgResponse (Maybe Text)

The label for the high-availability partition group.

describeHapgResponse_hsmsPendingRegistration :: Lens' DescribeHapgResponse (Maybe [Text])

describeHapgResponse_hsmsLastActionFailed :: Lens' DescribeHapgResponse (Maybe [Text])

describeHapgResponse_httpStatus :: Lens' DescribeHapgResponse Int

The response's http status code.