Network.AWS.CloudHSM.DescribeHsm

Description

This is documentation for AWS CloudHSM Classic. For more information, see AWS CloudHSM Classic FAQs, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

For information about the current version of AWS CloudHSM, see AWS CloudHSM, the AWS CloudHSM User Guide, and the AWS CloudHSM API Reference.

Retrieves information about an HSM. You can identify the HSM by its ARN or its serial number.

Creating a Request

data DescribeHsm

Contains the inputs for the DescribeHsm operation.

See: newDescribeHsm smart constructor.

Constructors

DescribeHsm'

Fields

hsmSerialNumber :: Maybe Text

The serial number of the HSM. Either the HsmArn or the HsmSerialNumber parameter must be specified.

hsmArn :: Maybe Text

The ARN of the HSM. Either the HsmArn or the SerialNumber parameter must be specified.

newDescribeHsm :: DescribeHsm

Create a value of DescribeHsm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsmSerialNumber:DescribeHsm', describeHsm_hsmSerialNumber - The serial number of the HSM. Either the HsmArn or the HsmSerialNumber parameter must be specified.

$sel:hsmArn:DescribeHsm', describeHsm_hsmArn - The ARN of the HSM. Either the HsmArn or the SerialNumber parameter must be specified.

Request Lenses

describeHsm_hsmSerialNumber :: Lens' DescribeHsm (Maybe Text)

The serial number of the HSM. Either the HsmArn or the HsmSerialNumber parameter must be specified.

describeHsm_hsmArn :: Lens' DescribeHsm (Maybe Text)

The ARN of the HSM. Either the HsmArn or the SerialNumber parameter must be specified.

Destructuring the Response

data DescribeHsmResponse

Contains the output of the DescribeHsm operation.

See: newDescribeHsmResponse smart constructor.

Constructors

DescribeHsmResponse'

Fields

subscriptionStartDate :: Maybe Text

The subscription start date.

status :: Maybe HsmStatus

The status of the HSM.

iamRoleArn :: Maybe Text

The ARN of the IAM role assigned to the HSM.

partitions :: Maybe [Text]

The list of partitions on the HSM.

statusDetails :: Maybe Text

Contains additional information about the status of the HSM.

subscriptionType :: Maybe SubscriptionType

eniIp :: Maybe Text

The IP address assigned to the HSM's ENI.

serverCertLastUpdated :: Maybe Text

The date and time that the server certificate was last updated.

eniId :: Maybe Text

The identifier of the elastic network interface (ENI) attached to the HSM.

availabilityZone :: Maybe Text

The Availability Zone that the HSM is in.

subnetId :: Maybe Text

The identifier of the subnet that the HSM is in.

hsmType :: Maybe Text

The HSM model type.

sshPublicKey :: Maybe Text

The public SSH key.

hsmArn :: Maybe Text

The ARN of the HSM.

serverCertUri :: Maybe Text

The URI of the certificate server.

serialNumber :: Maybe Text

The serial number of the HSM.

vendorName :: Maybe Text

The name of the HSM vendor.

subscriptionEndDate :: Maybe Text

The subscription end date.

vpcId :: Maybe Text

The identifier of the VPC that the HSM is in.

sshKeyLastUpdated :: Maybe Text

The date and time that the SSH key was last updated.

softwareVersion :: Maybe Text

The HSM software version.

httpStatus :: Int

The response's http status code.

newDescribeHsmResponse

Arguments

:: Int
$sel:httpStatus:DescribeHsmResponse'

-> DescribeHsmResponse

Create a value of DescribeHsmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionStartDate:DescribeHsmResponse', describeHsmResponse_subscriptionStartDate - The subscription start date.

$sel:status:DescribeHsmResponse', describeHsmResponse_status - The status of the HSM.

$sel:iamRoleArn:DescribeHsmResponse', describeHsmResponse_iamRoleArn - The ARN of the IAM role assigned to the HSM.

$sel:partitions:DescribeHsmResponse', describeHsmResponse_partitions - The list of partitions on the HSM.

$sel:statusDetails:DescribeHsmResponse', describeHsmResponse_statusDetails - Contains additional information about the status of the HSM.

$sel:subscriptionType:DescribeHsmResponse', describeHsmResponse_subscriptionType - Undocumented member.

$sel:eniIp:DescribeHsmResponse', describeHsmResponse_eniIp - The IP address assigned to the HSM's ENI.

$sel:serverCertLastUpdated:DescribeHsmResponse', describeHsmResponse_serverCertLastUpdated - The date and time that the server certificate was last updated.

$sel:eniId:DescribeHsmResponse', describeHsmResponse_eniId - The identifier of the elastic network interface (ENI) attached to the HSM.

$sel:availabilityZone:DescribeHsmResponse', describeHsmResponse_availabilityZone - The Availability Zone that the HSM is in.

$sel:subnetId:DescribeHsmResponse', describeHsmResponse_subnetId - The identifier of the subnet that the HSM is in.

$sel:hsmType:DescribeHsmResponse', describeHsmResponse_hsmType - The HSM model type.

$sel:sshPublicKey:DescribeHsmResponse', describeHsmResponse_sshPublicKey - The public SSH key.

$sel:hsmArn:DescribeHsm', describeHsmResponse_hsmArn - The ARN of the HSM.

$sel:serverCertUri:DescribeHsmResponse', describeHsmResponse_serverCertUri - The URI of the certificate server.

$sel:serialNumber:DescribeHsmResponse', describeHsmResponse_serialNumber - The serial number of the HSM.

$sel:vendorName:DescribeHsmResponse', describeHsmResponse_vendorName - The name of the HSM vendor.

$sel:subscriptionEndDate:DescribeHsmResponse', describeHsmResponse_subscriptionEndDate - The subscription end date.

$sel:vpcId:DescribeHsmResponse', describeHsmResponse_vpcId - The identifier of the VPC that the HSM is in.

$sel:sshKeyLastUpdated:DescribeHsmResponse', describeHsmResponse_sshKeyLastUpdated - The date and time that the SSH key was last updated.

$sel:softwareVersion:DescribeHsmResponse', describeHsmResponse_softwareVersion - The HSM software version.

$sel:httpStatus:DescribeHsmResponse', describeHsmResponse_httpStatus - The response's http status code.

Response Lenses

describeHsmResponse_subscriptionStartDate :: Lens' DescribeHsmResponse (Maybe Text)

The subscription start date.

describeHsmResponse_status :: Lens' DescribeHsmResponse (Maybe HsmStatus)

The status of the HSM.

describeHsmResponse_iamRoleArn :: Lens' DescribeHsmResponse (Maybe Text)

The ARN of the IAM role assigned to the HSM.

describeHsmResponse_partitions :: Lens' DescribeHsmResponse (Maybe [Text])

The list of partitions on the HSM.

describeHsmResponse_statusDetails :: Lens' DescribeHsmResponse (Maybe Text)

Contains additional information about the status of the HSM.

describeHsmResponse_subscriptionType :: Lens' DescribeHsmResponse (Maybe SubscriptionType)

Undocumented member.

describeHsmResponse_eniIp :: Lens' DescribeHsmResponse (Maybe Text)

The IP address assigned to the HSM's ENI.

describeHsmResponse_serverCertLastUpdated :: Lens' DescribeHsmResponse (Maybe Text)

The date and time that the server certificate was last updated.

describeHsmResponse_eniId :: Lens' DescribeHsmResponse (Maybe Text)

The identifier of the elastic network interface (ENI) attached to the HSM.

describeHsmResponse_availabilityZone :: Lens' DescribeHsmResponse (Maybe Text)

The Availability Zone that the HSM is in.

describeHsmResponse_subnetId :: Lens' DescribeHsmResponse (Maybe Text)

The identifier of the subnet that the HSM is in.

describeHsmResponse_hsmType :: Lens' DescribeHsmResponse (Maybe Text)

The HSM model type.

describeHsmResponse_sshPublicKey :: Lens' DescribeHsmResponse (Maybe Text)

The public SSH key.

describeHsmResponse_hsmArn :: Lens' DescribeHsmResponse (Maybe Text)

The ARN of the HSM.

describeHsmResponse_serverCertUri :: Lens' DescribeHsmResponse (Maybe Text)

The URI of the certificate server.

describeHsmResponse_serialNumber :: Lens' DescribeHsmResponse (Maybe Text)

The serial number of the HSM.

describeHsmResponse_vendorName :: Lens' DescribeHsmResponse (Maybe Text)

The name of the HSM vendor.

describeHsmResponse_subscriptionEndDate :: Lens' DescribeHsmResponse (Maybe Text)

The subscription end date.

describeHsmResponse_vpcId :: Lens' DescribeHsmResponse (Maybe Text)

The identifier of the VPC that the HSM is in.

describeHsmResponse_sshKeyLastUpdated :: Lens' DescribeHsmResponse (Maybe Text)

The date and time that the SSH key was last updated.

describeHsmResponse_softwareVersion :: Lens' DescribeHsmResponse (Maybe Text)

The HSM software version.

describeHsmResponse_httpStatus :: Lens' DescribeHsmResponse Int

The response's http status code.