Network.AWS.CloudHSM.GetConfig

Description

This is documentation for AWS CloudHSM Classic. For more information, see AWS CloudHSM Classic FAQs, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

For information about the current version of AWS CloudHSM, see AWS CloudHSM, the AWS CloudHSM User Guide, and the AWS CloudHSM API Reference.

Gets the configuration files necessary to connect to all high availability partition groups the client is associated with.

Creating a Request

data GetConfig

See: newGetConfig smart constructor.

Constructors

GetConfig'

Fields

clientArn :: Text

The ARN of the client.

clientVersion :: ClientVersion

The client version.

hapgList :: [Text]

A list of ARNs that identify the high-availability partition groups that are associated with the client.

newGetConfig

Arguments

:: Text
$sel:clientArn:GetConfig'

-> ClientVersion
$sel:clientVersion:GetConfig'

-> GetConfig

Create a value of GetConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientArn:GetConfig', getConfig_clientArn - The ARN of the client.

$sel:clientVersion:GetConfig', getConfig_clientVersion - The client version.

$sel:hapgList:GetConfig', getConfig_hapgList - A list of ARNs that identify the high-availability partition groups that are associated with the client.

Request Lenses

getConfig_clientArn :: Lens' GetConfig Text

The ARN of the client.

getConfig_clientVersion :: Lens' GetConfig ClientVersion

The client version.

getConfig_hapgList :: Lens' GetConfig [Text]

A list of ARNs that identify the high-availability partition groups that are associated with the client.

Destructuring the Response

data GetConfigResponse

See: newGetConfigResponse smart constructor.

Constructors

GetConfigResponse'

Fields

configFile :: Maybe Text

The chrystoki.conf configuration file.

configCred :: Maybe Text

The certificate file containing the server.pem files of the HSMs.

configType :: Maybe Text

The type of credentials.

httpStatus :: Int

The response's http status code.

newGetConfigResponse

Arguments

:: Int
$sel:httpStatus:GetConfigResponse'

-> GetConfigResponse

Create a value of GetConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configFile:GetConfigResponse', getConfigResponse_configFile - The chrystoki.conf configuration file.

$sel:configCred:GetConfigResponse', getConfigResponse_configCred - The certificate file containing the server.pem files of the HSMs.

$sel:configType:GetConfigResponse', getConfigResponse_configType - The type of credentials.

$sel:httpStatus:GetConfigResponse', getConfigResponse_httpStatus - The response's http status code.

Response Lenses

getConfigResponse_configFile :: Lens' GetConfigResponse (Maybe Text)

The chrystoki.conf configuration file.

getConfigResponse_configCred :: Lens' GetConfigResponse (Maybe Text)

The certificate file containing the server.pem files of the HSMs.

getConfigResponse_configType :: Lens' GetConfigResponse (Maybe Text)

The type of credentials.

getConfigResponse_httpStatus :: Lens' GetConfigResponse Int

The response's http status code.