Network.AWS.CloudHSM.Types

Description

Service Configuration

defaultService :: Service

API version 2014-05-30 of the Amazon CloudHSM SDK configuration.

Errors

_CloudHsmInternalException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that an internal error occurred.

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that one or more of the request parameters are not valid.

_CloudHsmServiceException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that an exception occurred in the AWS CloudHSM service.

ClientVersion

newtype ClientVersion

Constructors

ClientVersion'

Fields

fromClientVersion :: Text

Bundled Patterns

pattern ClientVersion_ :: ClientVersion

CloudHsmObjectState

newtype CloudHsmObjectState

Constructors

CloudHsmObjectState'

Fields

fromCloudHsmObjectState :: Text

Bundled Patterns

pattern CloudHsmObjectState_DEGRADED :: CloudHsmObjectState

pattern CloudHsmObjectState_READY :: CloudHsmObjectState

pattern CloudHsmObjectState_UPDATING :: CloudHsmObjectState

HsmStatus

newtype HsmStatus

Constructors

HsmStatus'

Fields

fromHsmStatus :: Text

Bundled Patterns

pattern HsmStatus_DEGRADED :: HsmStatus

pattern HsmStatus_PENDING :: HsmStatus

pattern HsmStatus_RUNNING :: HsmStatus

pattern HsmStatus_SUSPENDED :: HsmStatus

pattern HsmStatus_TERMINATED :: HsmStatus

pattern HsmStatus_TERMINATING :: HsmStatus

pattern HsmStatus_UPDATING :: HsmStatus

SubscriptionType

newtype SubscriptionType

Specifies the type of subscription for the HSM.

PRODUCTION - The HSM is being used in a production environment.

TRIAL - The HSM is being used in a product trial.

Constructors

SubscriptionType'

Fields

fromSubscriptionType :: Text

Bundled Patterns

pattern SubscriptionType_PRODUCTION :: SubscriptionType

Tag

data Tag

A key-value pair that identifies or specifies metadata about an AWS CloudHSM resource.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The key of the tag.

value :: Text

The value of the tag.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key of the tag.

$sel:value:Tag', tag_value - The value of the tag.

tag_key :: Lens' Tag Text

The key of the tag.

tag_value :: Lens' Tag Text

The value of the tag.