Network.AWS.CloudHSMV2.CreateCluster

Description

Creates a new AWS CloudHSM cluster.

Creating a Request

data CreateCluster

See: newCreateCluster smart constructor.

Constructors

CreateCluster'

Fields

backupRetentionPolicy :: Maybe BackupRetentionPolicy

A policy that defines how the service retains backups.

tagList :: Maybe [Tag]

Tags to apply to the CloudHSM cluster during creation.

sourceBackupId :: Maybe Text

The identifier (ID) of the cluster backup to restore. Use this value to restore the cluster from a backup instead of creating a new cluster. To find the backup ID, use DescribeBackups.

hsmType :: Text

The type of HSM to use in the cluster. Currently the only allowed value is hsm1.medium.

subnetIds :: NonEmpty Text

The identifiers (IDs) of the subnets where you are creating the cluster. You must specify at least one subnet. If you specify multiple subnets, they must meet the following criteria:

All subnets must be in the same virtual private cloud (VPC).

You can specify only one subnet per Availability Zone.

newCreateCluster

Arguments

:: Text
$sel:hsmType:CreateCluster'

-> NonEmpty Text
$sel:subnetIds:CreateCluster'

-> CreateCluster

Create a value of CreateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupRetentionPolicy:CreateCluster', createCluster_backupRetentionPolicy - A policy that defines how the service retains backups.

$sel:tagList:CreateCluster', createCluster_tagList - Tags to apply to the CloudHSM cluster during creation.

$sel:sourceBackupId:CreateCluster', createCluster_sourceBackupId - The identifier (ID) of the cluster backup to restore. Use this value to restore the cluster from a backup instead of creating a new cluster. To find the backup ID, use DescribeBackups.

$sel:hsmType:CreateCluster', createCluster_hsmType - The type of HSM to use in the cluster. Currently the only allowed value is hsm1.medium.

$sel:subnetIds:CreateCluster', createCluster_subnetIds - The identifiers (IDs) of the subnets where you are creating the cluster. You must specify at least one subnet. If you specify multiple subnets, they must meet the following criteria:

All subnets must be in the same virtual private cloud (VPC).

You can specify only one subnet per Availability Zone.

Request Lenses

createCluster_backupRetentionPolicy :: Lens' CreateCluster (Maybe BackupRetentionPolicy)

A policy that defines how the service retains backups.

createCluster_tagList :: Lens' CreateCluster (Maybe [Tag])

Tags to apply to the CloudHSM cluster during creation.

createCluster_sourceBackupId :: Lens' CreateCluster (Maybe Text)

The identifier (ID) of the cluster backup to restore. Use this value to restore the cluster from a backup instead of creating a new cluster. To find the backup ID, use DescribeBackups.

createCluster_hsmType :: Lens' CreateCluster Text

The type of HSM to use in the cluster. Currently the only allowed value is hsm1.medium.

createCluster_subnetIds :: Lens' CreateCluster (NonEmpty Text)

The identifiers (IDs) of the subnets where you are creating the cluster. You must specify at least one subnet. If you specify multiple subnets, they must meet the following criteria:

All subnets must be in the same virtual private cloud (VPC).

You can specify only one subnet per Availability Zone.

Destructuring the Response

data CreateClusterResponse

See: newCreateClusterResponse smart constructor.

Constructors

CreateClusterResponse'

Fields

cluster :: Maybe Cluster

Information about the cluster that was created.

httpStatus :: Int

The response's http status code.

newCreateClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateClusterResponse'

-> CreateClusterResponse

Create a value of CreateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:CreateClusterResponse', createClusterResponse_cluster - Information about the cluster that was created.

$sel:httpStatus:CreateClusterResponse', createClusterResponse_httpStatus - The response's http status code.

Response Lenses

createClusterResponse_cluster :: Lens' CreateClusterResponse (Maybe Cluster)

Information about the cluster that was created.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.